all

all

# S3 method for tensor
all(x, dim = -1, keepdim = FALSE, na.rm = FALSE)

Arguments

x
tensor object

dim
if negative (the default) will reduce to a scalar. Otherwise it will return TRUE if all elements in each row of the tensor in the given dimension `dim` are TRUE, FALSE otherwise.

keepdim
If keepdim is TRUE, the output tensor is of the same size as input except in the dimension dim where it is of size 1. Otherwise, dim is squeezed [squeeze()], resulting in the output tensor having 1 fewer dimension than input.

na.rm
won't be used by the function. Only there to be compatible with [all] generic.

Examples

x <- tensor(array(c(TRUE, FALSE, TRUE, TRUE), dim = c(2, 2)))
all(x)

#> tensor 
#> 0
#> [ Variable[CPUByteType]{} ]

all(x, dim = 1)

#> tensor 
#>  1
#>  0
#> [ Variable[CPUByteType]{2} ]

all(x, dim = 1, keepdim = FALSE)

#> tensor 
#>  1
#>  0
#> [ Variable[CPUByteType]{2} ]

all(x, dim = 2)

#> Error in tensor_all_(self$pointer, dim, keepdim): Dimension out of range (expected to be in range of [-2, 1], but got 2) (maybe_wrap_dim at /Users/administrator/nightlies/2018_10_05/wheel_build_dirs/libtorch_2.7/pytorch/aten/src/ATen/core/WrapDimMinimal.h:18)
#> frame #0: at::native::all_out(at::Tensor&, at::Tensor const&, long long, bool) + 128 (0x11a2c0e60 in libcaffe2.dylib)
#> frame #1: at::native::all(at::Tensor const&, long long, bool) + 133 (0x11a2c0cd5 in libcaffe2.dylib)
#> frame #2: at::TypeDefault::all(at::Tensor const&, long long, bool) const + 123 (0x11a50a21b in libcaffe2.dylib)
#> frame #3: torch::autograd::VariableType::all(at::Tensor const&, long long, bool) const + 288 (0x1150b6ad0 in libtorch.1.dylib)
#> frame #4: tensor_all_(Rcpp::XPtr<at::Tensor, Rcpp::PreserveStorage, &(void Rcpp::standard_delete_finalizer<at::Tensor>(at::Tensor*)), false>, long long, bool) + 91 (0x115d5a52b in torch.so)
#> frame #5: _torch_tensor_all_ + 115 (0x115d4f753 in torch.so)
#> frame #6: R_doDotCall + 979 (0x110d454b3 in libR.dylib)
#> frame #7: do_dotcall + 323 (0x110d47333 in libR.dylib)
#> frame #8: Rf_eval + 1658 (0x110d75c9a in libR.dylib)
#> frame #9: do_begin + 432 (0x110d8a580 in libR.dylib)
#> frame #10: Rf_eval + 942 (0x110d759ce in libR.dylib)
#> frame #11: R_execClosure + 3004 (0x110d8814c in libR.dylib)
#> frame #12: Rf_eval + 1772 (0x110d75d0c in libR.dylib)
#> frame #13: forcePromise + 179 (0x110d86513 in libR.dylib)
#> frame #14: Rf_eval + 1182 (0x110d75abe in libR.dylib)
#> frame #15: forcePromise + 179 (0x110d86513 in libR.dylib)
#> frame #16: Rf_eval + 1182 (0x110d75abe in libR.dylib)
#> frame #17: forcePromise + 179 (0x110d86513 in libR.dylib)
#> frame #18: Rf_eval + 1182 (0x110d75abe in libR.dylib)
#> frame #19: do_set + 155 (0x110d8a96b in libR.dylib)
#> frame #20: Rf_eval + 942 (0x110d759ce in libR.dylib)
#> frame #21: do_begin + 432 (0x110d8a580 in libR.dylib)
#> frame #22: Rf_eval + 942 (0x110d759ce in libR.dylib)
#> frame #23: R_execClosure + 3004 (0x110d8814c in libR.dylib)
#> frame #24: Rf_eval + 1772 (0x110d75d0c in libR.dylib)
#> frame #25: do_begin + 432 (0x110d8a580 in libR.dylib)
#> frame #26: Rf_eval + 942 (0x110d759ce in libR.dylib)
#> frame #27: Rf_eval + 942 (0x110d759ce in libR.dylib)
#> frame #28: do_begin + 432 (0x110d8a580 in libR.dylib)
#> frame #29: Rf_eval + 942 (0x110d759ce in libR.dylib)
#> frame #30: R_execClosure + 3004 (0x110d8814c in libR.dylib)
#> frame #31: Rf_eval + 1772 (0x110d75d0c in libR.dylib)
#> frame #32: do_begin + 432 (0x110d8a580 in libR.dylib)
#> frame #33: Rf_eval + 942 (0x110d759ce in libR.dylib)
#> frame #34: R_execClosure + 3004 (0x110d8814c in libR.dylib)
#> frame #35: Rf_eval + 1772 (0x110d75d0c in libR.dylib)
#> frame #36: do_begin + 432 (0x110d8a580 in libR.dylib)
#> frame #37: Rf_eval + 942 (0x110d759ce in libR.dylib)
#> frame #38: R_execClosure + 3004 (0x110d8814c in libR.dylib)
#> frame #39: Rf_eval + 1772 (0x110d75d0c in libR.dylib)
#> frame #40: do_begin + 432 (0x110d8a580 in libR.dylib)
#> frame #41: Rf_eval + 942 (0x110d759ce in libR.dylib)
#> frame #42: R_execClosure + 3004 (0x110d8814c in libR.dylib)
#> frame #43: Rf_eval + 1772 (0x110d75d0c in libR.dylib)
#> frame #44: do_begin + 432 (0x110d8a580 in libR.dylib)
#> frame #45: Rf_eval + 942 (0x110d759ce in libR.dylib)
#> frame #46: R_execClosure + 3004 (0x110d8814c in libR.dylib)
#> frame #47: Rf_DispatchGroup + 2160 (0x110d8d160 in libR.dylib)
#> frame #48: do_logic3 + 196 (0x110db84f4 in libR.dylib)
#> frame #49: Rf_eval + 1811 (0x110d75d33 in libR.dylib)
#> frame #50: do_eval + 645 (0x110d8b8f5 in libR.dylib)
#> frame #51: bcEval + 19608 (0x110d7aba8 in libR.dylib)
#> frame #52: Rf_eval + 443 (0x110d757db in libR.dylib)
#> frame #53: R_execClosure + 3004 (0x110d8814c in libR.dylib)
#> frame #54: bcEval + 18659 (0x110d7a7f3 in libR.dylib)
#> frame #55: Rf_eval + 443 (0x110d757db in libR.dylib)
#> frame #56: forcePromise + 179 (0x110d86513 in libR.dylib)
#> frame #57: Rf_eval + 1182 (0x110d75abe in libR.dylib)
#> frame #58: do_withVisible + 49 (0x110d8bf61 in libR.dylib)
#> frame #59: do_internal + 339 (0x110dc9393 in libR.dylib)
#> frame #60: bcEval + 20121 (0x110d7ada9 in libR.dylib)
#> frame #61: Rf_eval + 443 (0x110d757db in libR.dylib)
#> frame #62: R_execClosure + 3004 (0x110d8814c in libR.dylib)
#> frame #63: bcEval + 18659 (0x110d7a7f3 in libR.dylib)