Computes the eigenvalue decomposition of a complex Hermitian or real symmetric matrix.

Letting \(\mathbb{K}\) be \(\mathbb{R}\) or \(\mathbb{C}\), the eigenvalue decomposition of a complex Hermitian or real symmetric matrix \(A \in \mathbb{K}^{n \times n}\) is defined as

linalg_eigh(A, UPLO = "L")

Arguments

A
(Tensor): tensor of shape (*, n, n) where * is zero or more batch dimensions consisting of symmetric or Hermitian matrices.

UPLO
('L', 'U', optional): controls whether to use the upper or lower triangular part of A in the computations. Default: 'L'.

Value

A list (eigenvalues, eigenvectors) which corresponds to \(\Lambda\) and \(Q\) above. eigenvalues will always be real-valued, even when A is complex.

It will also be ordered in ascending order. eigenvectors will have the same dtype as A and will contain the eigenvectors as its columns.

Details

A=Qdiag⁡(Λ)QHQ∈Kn×n,Λ∈Rn A = Q \operatorname{diag}(\Lambda) Q^{H}\mathrlap{\qquad Q \in \mathbb{K}^{n \times n}, \Lambda \in \mathbb{R}^n} A=Qdiag(Λ)QHQ∈Kn×n,Λ∈Rn

where \(Q^{H}\) is the conjugate transpose when \(Q\) is complex, and the transpose when \(Q\) is real-valued. \(Q\) is orthogonal in the real case and unitary in the complex case.

Supports input of float, double, cfloat and cdouble dtypes. Also supports batches of matrices, and if A is a batch of matrices then the output has the same batch dimensions.

A is assumed to be Hermitian (resp. symmetric), but this is not checked internally, instead:

If UPLO\ = 'L' (default), only the lower triangular part of the matrix is used in the computation.

If UPLO\ = 'U', only the upper triangular part of the matrix is used. The eigenvalues are returned in ascending order.

Note

The eigenvalues of real symmetric or complex Hermitian matrices are always real.

Warning

The eigenvectors of a symmetric matrix are not unique, nor are they continuous with respect to A. Due to this lack of uniqueness, different hardware and software may compute different eigenvectors. This non-uniqueness is caused by the fact that multiplying an eigenvector by -1 in the real case or by \(e^{i \phi}, \phi \in \mathbb{R}\) in the complex case produces another set of valid eigenvectors of the matrix. This non-uniqueness problem is even worse when the matrix has repeated eigenvalues. In this case, one may multiply the associated eigenvectors spanning the subspace by a rotation matrix and the resulting eigenvectors will be valid eigenvectors.

Gradients computed using the eigenvectors tensor will only be finite when A has unique eigenvalues. Furthermore, if the distance between any two eigvalues is close to zero, the gradient will be numerically unstable, as it depends on the eigenvalues \(\lambda_i\) through the computation of \(\frac{1}{\min_{i \neq j} \lambda_i - \lambda_j}\).

See also

Examples

if (torch_is_installed()) {
a <- torch_randn(2, 2)
linalg_eigh(a)

}

#> [[1]]
#> torch_tensor
#>  0.0487
#>  1.4760
#> [ CPUFloatType{2} ]
#> 
#> [[2]]
#> torch_tensor
#> -0.9008 -0.4343
#> -0.4343  0.9008
#> [ CPUFloatType{2,2} ]
#>