Computes the eigenvalues of a square matrix.

Letting be or , the eigenvalues of a square matrix are defined as the roots (counted with multiplicity) of the polynomial p of degree n given by

linalg_eigvals(A)

Arguments

A

(Tensor): tensor of shape (*, n, n) where * is zero or more batch dimensions.

Details

p(λ)=det⁡(A−λIn)λ∈C p(\lambda) = \operatorname{det}(A - \lambda \mathrm{I}_n)\mathrlap{\qquad \lambda \in \mathbb{C}} p(λ)=det(A−λIn​)λ∈C

where is the n-dimensional identity matrix. Supports input of float, double, cfloat and cdouble dtypes. Also supports batches of matrices, and if A is a batch of matrices then the output has the same batch dimensions.

Note

The eigenvalues of a real matrix may be complex, as the roots of a real polynomial may be complex. The eigenvalues of a matrix are always well-defined, even when the matrix is not diagonalizable.

See also

Examples

if (torch_is_installed()) {
a <- torch_randn(2, 2)
w <- linalg_eigvals(a)
}