Computes the eigenvalues of a complex Hermitian or real symmetric matrix.

Letting be or , the eigenvalues of a complex Hermitian or real symmetric matrix are defined as the roots (counted with multiplicity) of the polynomial p of degree n given by

Usage

linalg_eigvalsh(A, UPLO = "L")

Arguments

A

(Tensor): tensor of shape (*, n, n) where * is zero or more batch dimensions consisting of symmetric or Hermitian matrices.

UPLO

('L', 'U', optional): controls whether to use the upper or lower triangular part of A in the computations. Default: 'L'.

Value

A real-valued tensor cointaining the eigenvalues even when A is complex. The eigenvalues are returned in ascending order.

Details

p(λ)=det⁡(A−λIn)λ∈R p(\lambda) = \operatorname{det}(A - \lambda \mathrm{I}_n)\mathrlap{\qquad \lambda \in \mathbb{R}} p(λ)=det(A−λIn​)λ∈R

where is the n-dimensional identity matrix.

The eigenvalues of a real symmetric or complex Hermitian matrix are always real. Supports input of float, double, cfloat and cdouble dtypes. Also supports batches of matrices, and if A is a batch of matrices then the output has the same batch dimensions. The eigenvalues are returned in ascending order.

A is assumed to be Hermitian (resp. symmetric), but this is not checked internally, instead:

If UPLO\ = 'L' (default), only the lower triangular part of the matrix is used in the computation.

If UPLO\ = 'U', only the upper triangular part of the matrix is used.

See also

Examples

if (torch_is_installed()) {
a <- torch_randn(2, 2)
linalg_eigvalsh(a)
}
#> torch_tensor
#> -1.8745
#>  0.0572
#> [ CPUFloatType{2} ]