Computes the inverse of a square matrix if it exists.

Throws a runtime_error if the matrix is not invertible.

Usage

linalg_inv(A)

Arguments

A

(Tensor): tensor of shape (*, n, n) where * is zero or more batch dimensions consisting of invertible matrices.

Details

Letting be or , for a matrix , its inverse matrix (if it exists) is defined as

A−1A=AA−1=In A^{-1}A = AA^{-1} = \mathrm{I}_n A−1A=AA−1=In​ where is the n-dimensional identity matrix.

The inverse matrix exists if and only if is invertible. In this case, the inverse is unique. Supports input of float, double, cfloat and cdouble dtypes. Also supports batches of matrices, and if A is a batch of matrices then the output has the same batch dimensions.

Consider using linalg_solve() if possible for multiplying a matrix on the left by the inverse, as linalg_solve(A, B) == A$inv() %*% B It is always prefered to use linalg_solve() when possible, as it is faster and more numerically stable than computing the inverse explicitly.

See also

Examples

if (torch_is_installed()) {
A <- torch_randn(4, 4)
linalg_inv(A)
}
#> torch_tensor
#> -1.1932  0.1808  0.4198  0.9346
#>  0.0076 -0.1740  0.1018  0.4838
#> -0.7966  0.4620 -0.1211  0.5590
#>  1.4917  0.3207  0.2693 -1.2210
#> [ CPUFloatType{4,4} ]