Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 1.4806e-01  1.2782e-01 -1.3181e+00 -1.6748e-01  1.5760e+00 -8.1000e-01
#> -1.2938e+00 -4.6650e-01 -2.3092e+00 -4.5060e-01  2.5774e-01  2.4798e+00
#>  1.5889e+00  5.6571e-01 -1.0141e+00 -2.3078e+00  1.5629e-01 -1.0670e+00
#>  9.2378e-01 -8.1461e-01 -4.6645e-01  1.3164e+00 -4.3752e-01 -4.4730e-01
#>  9.8056e-01  6.2474e-01 -8.2288e-01 -5.6566e-02  1.5581e-01 -3.0590e-01
#> -1.1376e+00 -1.6929e-02  6.5225e-01  1.2261e+00  3.4274e-02  6.0990e-01
#> -2.5636e-01 -1.0151e+00  1.0557e+00 -1.4429e+00 -4.9996e-01 -2.4900e-01
#> -5.2633e-01 -1.7527e+00  6.1899e-01 -4.0355e-02  9.1104e-01  1.0707e+00
#>  2.6833e-01  2.0920e-01 -1.1437e-01  6.7720e-01  3.5489e-01  1.9509e+00
#>  8.3883e-01 -5.3670e-01  6.7501e-01 -4.5682e-01  7.0336e-01 -1.1345e-01
#> -5.5784e-01  6.7038e-01 -2.5191e+00  8.9835e-01  1.3193e+00 -1.1451e-01
#> -3.3150e-01  4.9263e-01  8.2023e-01  5.2012e-01 -2.6908e-01 -3.1437e+00
#>  6.6839e-01 -2.7696e-01 -2.8709e-01 -5.6308e-01  5.2030e-01  1.5194e+00
#> -5.3008e-01 -1.7213e-01  6.8710e-01 -2.8405e+00  2.1675e-01 -2.5220e-01
#>  4.5881e-01 -3.3413e-01 -8.4972e-02  9.1190e-02 -1.2698e+00  1.2989e-01
#>  1.5626e+00 -4.2226e-01  2.2780e-01 -1.1271e+00  1.0882e+00 -6.5972e-02
#> -2.3835e-01  2.5675e-01 -5.6019e-01  7.3356e-01  9.2115e-01 -3.5177e-01
#>  7.0454e-01 -1.3372e+00 -7.4534e-01  1.0861e-01  3.7911e-01 -5.1990e-01
#> -7.5273e-02 -5.6726e-01 -3.3641e-01  6.9058e-01  6.8525e-02 -1.4442e+00
#> -1.9223e-01 -2.0833e-01  2.2206e+00 -9.1891e-01  3.1117e-01 -1.8709e-01
#> -2.0944e+00  2.2125e-01 -6.7557e-01  7.8282e-02  2.1208e-01 -2.6905e-01
#>  7.9249e-01 -8.2880e-02  9.1596e-01 -6.2696e-01  8.5045e-01 -8.8604e-01
#>  3.4774e-01 -6.1513e-01  4.2972e-01  1.2299e-01  1.5044e+00 -3.2752e-01
#> -4.2755e-01  6.8444e-01 -1.9903e+00  7.2690e-02 -6.7937e-01 -3.1411e-01
#>  7.5563e-01 -1.7070e-01  1.2023e+00 -6.5837e-01  3.7644e-01 -7.0506e-01
#>  1.0411e+00  9.8568e-01 -1.5579e+00 -6.9598e-01  6.0912e-02 -2.1489e-02
#> -5.0832e-01  8.5130e-01 -3.1047e-01 -7.6991e-01 -1.0782e+00 -5.9076e-01
#>  3.2108e-01 -4.3321e-01  8.6882e-01 -4.4918e-01 -1.5815e+00 -6.6611e-01
#> -5.4731e-01  1.5741e-02  1.3690e+00  1.2820e+00  7.5336e-01  2.9747e-01
#>  2.5130e-01 -6.1186e-01  1.6347e-01 -8.1403e-01  9.6848e-01  2.5667e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]