Kaiming normal initialization

Fills the input Tensor with values according to the method described in Delving deep into rectifiers: Surpassing human-level performance on ImageNet classification - He, K. et al. (2015), using a normal distribution.

nn_init_kaiming_normal_(
  tensor,
  a = 0,
  mode = "fan_in",
  nonlinearity = "leaky_relu"
)

Arguments

tensor
an n-dimensional torch.Tensor

a
the negative slope of the rectifier used after this layer (only used with 'leaky_relu')

mode
either 'fan_in' (default) or 'fan_out'. Choosing 'fan_in' preserves the magnitude of the variance of the weights in the forward pass. Choosing 'fan_out' preserves the magnitudes in the backwards pass.

nonlinearity
the non-linear function. recommended to use only with 'relu' or 'leaky_relu' (default).

Examples

w <- torch_empty(3, 5)
nn_init_kaiming_normal_(w, mode = "fan_in", nonlinearity = "leaky_relu")

#> torch_tensor 
#> -0.7415 -0.4026 -0.1830  0.9564  0.5827
#>  0.8907 -1.0440  0.0456 -0.1924  1.1919
#> -1.1077 -0.2889  0.0341  0.3169 -0.3735
#> [ CPUFloatType{3,5} ]