MaxPool2D module

Applies a 2D max pooling over an input signal composed of several input planes.

nn_max_pool2d(...)

Arguments

kernel_size
the size of the window to take a max over

stride
the stride of the window. Default value is kernel_size

padding
implicit zero padding to be added on both sides

dilation
a parameter that controls the stride of elements in the window

return_indices
if TRUE, will return the max indices along with the outputs. Useful for nn_max_unpool2d later.

ceil_mode
when TRUE, will use ceil instead of floor to compute the output shape

Details

In the simplest case, the output value of the layer with input size \((N, C, H, W)\), output \((N, C, H_{out}, W_{out})\) and kernel_size \((kH, kW)\) can be precisely described as:

$$ \begin{aligned} out(N_i, C_j, h, w) ={} & \max_{m=0, \ldots, kH-1} \max_{n=0, \ldots, kW-1} \\ & \text{input}(N_i, C_j, \text{stride[0]} \times h + m, \text{stride[1]} \times w + n) \end{aligned}$$

If padding is non-zero, then the input is implicitly zero-padded on both sides for padding number of points. dilation controls the spacing between the kernel points. It is harder to describe, but this link_ has a nice visualization of what dilation does.

The parameters kernel_size, stride, padding, dilation can either be:

a single int -- in which case the same value is used for the height and width dimension

a tuple of two ints -- in which case, the first int is used for the height dimension, and the second int for the width dimension

Shape

Input: \((N, C, H_{in}, W_{in})\)

Output: \((N, C, H_{out}, W_{out})\), where

$$ H_{out} = \left\lfloor\frac{H_{in} + 2 * \text{padding[0]} - \text{dilation[0]} \times (\text{kernel\_size[0]} - 1) - 1}{\text{stride[0]}} + 1\right\rfloor $$

$$ W_{out} = \left\lfloor\frac{W_{in} + 2 * \text{padding[1]} - \text{dilation[1]} \times (\text{kernel\_size[1]} - 1) - 1}{\text{stride[1]}} + 1\right\rfloor $$

Examples

# pool of square window of size=3, stride=2
m <- nn_max_pool2d(3, stride=2)
# pool of non-square window
m <- nn_max_pool2d(c(3, 2), stride=c(2, 1))
input <- torch_randn(20, 16, 50, 32)
output <- m(input)