Gels

Computes the solution to the least squares and least norm problems for a full rank matrix A of size (m,n) and a matrix B of size (m,k).

tch_gels(x, A)

Arguments

x
tensor object

A
the m,b by nn matrix A

Examples

A <- tensor(matrix(runif(100), ncol = 10))
y <- tensor(matrix(runif(10), ncol = 1))
tch_gels(y, A)

#> [[1]]
#> tensor 
#>  0.8667
#>  0.3988
#>  1.7198
#> -0.4065
#> -0.2945
#> -0.2274
#>  0.1860
#>  0.0667
#> -0.5807
#> -0.8533
#> [ Variable[CPUFloatType]{10,1} ]
#> 
#> [[2]]
#> tensor 
#> -1.4555 -1.4805 -1.0961 -1.5807 -0.8747 -1.3722 -1.2065 -0.8305 -1.4904 -0.6710
#>  0.1663 -1.3771 -1.1747 -0.3049 -1.0116 -0.2729 -0.7190 -0.8029 -1.1489 -0.8212
#>  0.1106  0.0725 -1.1713 -0.4670 -0.6837 -0.8388 -0.5508 -0.3651 -0.3438 -0.7331
#>  0.2003  0.3363 -0.0783  0.6128 -0.1225 -0.0644 -0.1414 -0.0799 -0.3100  0.1821
#>  0.0209  0.5067 -0.0875  0.3446 -0.7893 -0.2744  0.1276 -0.6421 -0.4239  0.1155
#>  0.0010  0.0525  0.4580 -0.2846  0.3881  0.5116  0.2312  0.1751  0.1206  0.0248
#>  0.2034  0.2244  0.0968 -0.1615  0.1027  0.5989  0.9356 -0.0773 -0.1048  0.3397
#>  0.1000  0.2309  0.2821 -0.4125 -0.5746  0.0929  0.3690 -0.4933 -0.1018 -0.3896
#>  0.1049  0.2997 -0.0652  0.1697 -0.2212  0.0576  0.0058 -0.0289 -0.6900  0.5166
#>  0.2266  0.1795 -0.1284 -0.2075  0.1706 -0.3391 -0.1203  0.4983  0.7645 -0.3664
#> [ Variable[CPUFloatType]{10,10} ]
#>