Add

Add

Arguments

input
(Tensor) the input tensor.

value
(Number) the number to be added to each element of input

other
(Tensor) the second input tensor

alpha
(Number) the scalar multiplier for other

add(input, other, out=None)

Adds the scalar other to each element of the input input and returns a new resulting tensor.

$$ \text{out} = \text{input} + \text{other} $$ If input is of type FloatTensor or DoubleTensor, other must be a real number, otherwise it should be an integer.

add(input, other, *, alpha=1, out=None)

Each element of the tensor other is multiplied by the scalar alpha and added to each element of the tensor input. The resulting tensor is returned.

The shapes of input and other must be broadcastable .

$$ \text{out} = \text{input} + \text{alpha} \times \text{other} $$ If other is of type FloatTensor or DoubleTensor, alpha must be a real number, otherwise it should be an integer.

Examples

a = torch_randn(c(4))
a

#> torch_tensor 
#>  0.3766
#>  0.2980
#> -0.3306
#> -0.1553
#> [ CPUFloatType{4} ]

torch_add(a, 20)

#> torch_tensor 
#>  20.3766
#>  20.2980
#>  19.6694
#>  19.8447
#> [ CPUFloatType{4} ]

a = torch_randn(c(4))
a

#> torch_tensor 
#>  0.2634
#> -0.3127
#>  0.3279
#>  0.0938
#> [ CPUFloatType{4} ]

b = torch_randn(c(4, 1))
b

#> torch_tensor 
#> -0.9428
#>  0.2656
#> -0.0625
#>  1.8104
#> [ CPUFloatType{4,1} ]

torch_add(a, b)

#> torch_tensor 
#> -0.6794 -1.2555 -0.6149 -0.8490
#>  0.5290 -0.0471  0.5935  0.3594
#>  0.2009 -0.3753  0.2654  0.0313
#>  2.0737  1.4976  2.1383  1.9042
#> [ CPUFloatType{4,4} ]