Addcmul

Addcmul

Arguments

input
(Tensor) the tensor to be added

tensor1
(Tensor) the tensor to be multiplied

tensor2
(Tensor) the tensor to be multiplied

value
(Number, optional) multiplier for \(tensor1 .* tensor2\)

out
(Tensor, optional) the output tensor.

addcmul(input, tensor1, tensor2, *, value=1, out=None) -> Tensor

Performs the element-wise multiplication of tensor1 by tensor2, multiply the result by the scalar value and add it to input.

$$ \text{out}_i = \text{input}_i + \text{value} \times \text{tensor1}_i \times \text{tensor2}_i $$ The shapes of tensor, tensor1, and tensor2 must be broadcastable .

For inputs of type FloatTensor or DoubleTensor, value must be a real number, otherwise an integer.

Examples

t = torch_randn(c(1, 3))
t1 = torch_randn(c(3, 1))
t2 = torch_randn(c(1, 3))
torch_addcmul(t, t1, t2, 0.1)

#> torch_tensor 
#> -0.8718 -0.1021  0.3403
#> -0.6687  0.0077  0.6375
#> -0.7381 -0.0298  0.5360
#> [ CPUFloatType{3,3} ]