Argmin

Argmin

Arguments

input
(Tensor) the input tensor.

dim
(int) the dimension to reduce. If None, the argmin of the flattened input is returned.

keepdim
(bool) whether the output tensor has dim retained or not. Ignored if dim=None.

argmin(input) -> LongTensor

Returns the indices of the minimum value of all elements in the input tensor.

This is the second value returned by torch_min. See its documentation for the exact semantics of this method.

argmin(input, dim, keepdim=False, out=None) -> LongTensor

Returns the indices of the minimum values of a tensor across a dimension.

This is the second value returned by torch_min. See its documentation for the exact semantics of this method.

Examples

a = torch_randn(c(4, 4))
a

#> torch_tensor 
#>  1.6416  1.0551 -0.1648 -1.7722
#>  1.7110 -0.4657  0.0843  1.5698
#> -0.9455  1.3457  0.2076 -0.9894
#> -2.0847  1.1386 -0.2182  1.0545
#> [ CPUFloatType{4,4} ]

torch_argmin(a)

#> torch_tensor 
#> 12
#> [ CPULongType{} ]

a = torch_randn(c(4, 4))
a

#> torch_tensor 
#> -0.3604 -1.0423  0.9780 -1.2052
#>  0.3993 -0.0517 -0.7131 -1.4085
#> -0.1945  0.3397 -1.2391 -0.1698
#> -0.9233 -0.2634  1.4732 -0.1863
#> [ CPUFloatType{4,4} ]

torch_argmin(a, dim=1)

#> torch_tensor 
#>  3
#>  3
#>  2
#>  0
#> [ CPULongType{4} ]