Cholesky

Cholesky

Arguments

input
(Tensor) the input tensor \(A\) of size \((*, n, n)\) where * is zero or more batch dimensions consisting of symmetric positive-definite matrices.

upper
(bool, optional) flag that indicates whether to return a upper or lower triangular matrix. Default: False

out
(Tensor, optional) the output matrix

cholesky(input, upper=False, out=None) -> Tensor

Computes the Cholesky decomposition of a symmetric positive-definite matrix \(A\) or for batches of symmetric positive-definite matrices.

If upper is True, the returned matrix U is upper-triangular, and the decomposition has the form:

$$ A = U^TU $$ If upper is False, the returned matrix L is lower-triangular, and the decomposition has the form:

$$ A = LL^T $$ If upper is True, and \(A\) is a batch of symmetric positive-definite matrices, then the returned tensor will be composed of upper-triangular Cholesky factors of each of the individual matrices. Similarly, when upper is False, the returned tensor will be composed of lower-triangular Cholesky factors of each of the individual matrices.

Examples

a = torch_randn(c(3, 3))
a = torch_mm(a, a$t()) # make symmetric positive-definite
l = torch_cholesky(a)
a

#> torch_tensor 
#>  7.1574 -0.3896  4.2226
#> -0.3896  1.6170  1.7057
#>  4.2226  1.7057  8.0317
#> [ CPUFloatType{3,3} ]

l

#> torch_tensor 
#>  2.6753  0.0000  0.0000
#> -0.1456  1.2632  0.0000
#>  1.5783  1.5322  1.7868
#> [ CPUFloatType{3,3} ]

torch_mm(l, l$t())

#> torch_tensor 
#>  7.1574 -0.3896  4.2226
#> -0.3896  1.6170  1.7057
#>  4.2226  1.7057  8.0317
#> [ CPUFloatType{3,3} ]

a = torch_randn(c(3, 2, 2))
if (FALSE) {
a = torch_matmul(a, a$transpose(-1, -2)) + 1e-03 # make symmetric positive-definite
l = torch_cholesky(a)
z = torch_matmul(l, l$transpose(-1, -2))
torch_max(torch_abs(z - a)) # Max non-zero
}