Cholesky_inverse

Cholesky_inverse

Arguments

input
(Tensor) the input 2-D tensor \(u\), a upper or lower triangular Cholesky factor

upper
(bool, optional) whether to return a lower (default) or upper triangular matrix

out
(Tensor, optional) the output tensor for inv

cholesky_inverse(input, upper=False, out=None) -> Tensor

Computes the inverse of a symmetric positive-definite matrix \(A\) using its Cholesky factor \(u\): returns matrix inv. The inverse is computed using LAPACK routines dpotri and spotri (and the corresponding MAGMA routines).

If upper is False, \(u\) is lower triangular such that the returned tensor is

$$ inv = (uu^{{T}})^{{-1}} $$ If upper is True or not provided, \(u\) is upper triangular such that the returned tensor is

$$ inv = (u^T u)^{{-1}} $$

Examples

if (FALSE) {
a = torch_randn(c(3, 3))
a = torch_mm(a, a$t()) + 1e-05 * torch_eye(3) # make symmetric positive definite
u = torch_cholesky(a)
a
torch_cholesky_inverse(u)
a$inverse()
}