Cholesky_solve

Cholesky_solve

Usage

torch_cholesky_solve(self, input2, upper = FALSE)

Arguments

self

(Tensor) input matrix \(b\) of size \((*, m, k)\), where \(*\) is zero or more batch dimensions

input2

(Tensor) input matrix \(u\) of size \((*, m, m)\), where \(*\) is zero of more batch dimensions composed of upper or lower triangular Cholesky factor

upper

(bool, optional) whether to consider the Cholesky factor as a lower or upper triangular matrix. Default: FALSE.

cholesky_solve(input, input2, upper=False, out=NULL) -> Tensor

Solves a linear system of equations with a positive semidefinite matrix to be inverted given its Cholesky factor matrix \(u\).

If upper is FALSE, \(u\) is and lower triangular and c is returned such that:

$$ c = (u u^T)^{{-1}} b $$ If upper is TRUE or not provided, \(u\) is upper triangular and c is returned such that:

$$ c = (u^T u)^{{-1}} b $$ torch_cholesky_solve(b, u) can take in 2D inputs b, u or inputs that are batches of 2D matrices. If the inputs are batches, then returns batched outputs c

Examples

if (torch_is_installed()) {

a = torch_randn(c(3, 3))
a = torch_mm(a, a$t()) # make symmetric positive definite
u = torch_cholesky(a)
a
b = torch_randn(c(3, 2))
b
torch_cholesky_solve(b, u)
torch_mm(a$inverse(), b)
}
#> Error in (function (self, upper) {    .Call(`_torch_cpp_torch_namespace_cholesky_self_Tensor`,         self, upper)})(self = <pointer: 0x600001faf3a0>, upper = FALSE): cholesky: The factorization could not be completed because the input is not positive-definite (the leading minor of order 3 is not positive-definite).
#> Exception raised from _linalg_check_errors at /Users/runner/work/libtorch-mac-m1/libtorch-mac-m1/pytorch/aten/src/ATen/native/BatchLinearAlgebra.cpp:1551 (most recent call first):
#> frame #0: std::__1::shared_ptr<c10::(anonymous namespace)::PyTorchStyleBacktrace> std::__1::make_shared[abi:ue170006]<c10::(anonymous namespace)::PyTorchStyleBacktrace, c10::SourceLocation&, void>(c10::SourceLocation&) + 121 (0x104d1d639 in libc10.dylib)
#> frame #1: c10::Error::Error(c10::SourceLocation, std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>) + 54 (0x104d1d4f6 in libc10.dylib)
#> frame #2: at::native::_linalg_check_errors(at::Tensor const&, c10::basic_string_view<char>, bool) + 1924 (0x1171497c4 in libtorch_cpu.dylib)
#> frame #3: at::_ops::_linalg_check_errors::call(at::Tensor const&, c10::basic_string_view<char>, bool) + 306 (0x117f393d2 in libtorch_cpu.dylib)
#> frame #4: at::native::cholesky(at::Tensor const&, bool) + 733 (0x11714c84d in libtorch_cpu.dylib)
#> frame #5: c10::impl::wrap_kernel_functor_unboxed_<c10::impl::detail::WrapFunctionIntoFunctor_<c10::CompileTimeFunctionPointer<at::Tensor (at::Tensor const&, bool), &at::(anonymous namespace)::(anonymous namespace)::wrapper_CPU__cholesky(at::Tensor const&, bool)>, at::Tensor, c10::guts::typelist::typelist<at::Tensor const&, bool>>, at::Tensor (at::Tensor const&, bool)>::call(c10::OperatorKernel*, c10::DispatchKeySet, at::Tensor const&, bool) + 20 (0x118255474 in libtorch_cpu.dylib)
#> frame #6: at::_ops::cholesky::redispatch(c10::DispatchKeySet, at::Tensor const&, bool) + 95 (0x117a6a8cf in libtorch_cpu.dylib)
#> frame #7: c10::impl::wrap_kernel_functor_unboxed_<c10::impl::detail::WrapFunctionIntoFunctor_<c10::CompileTimeFunctionPointer<at::Tensor (c10::DispatchKeySet, at::Tensor const&, bool), &torch::autograd::VariableType::(anonymous namespace)::cholesky(c10::DispatchKeySet, at::Tensor const&, bool)>, at::Tensor, c10::guts::typelist::typelist<c10::DispatchKeySet, at::Tensor const&, bool>>, at::Tensor (c10::DispatchKeySet, at::Tensor const&, bool)>::call(c10::OperatorKernel*, c10::DispatchKeySet, at::Tensor const&, bool) + 762 (0x11a0f281a in libtorch_cpu.dylib)
#> frame #8: at::_ops::cholesky::call(at::Tensor const&, bool) + 286 (0x117a6a63e in libtorch_cpu.dylib)
#> frame #9: at::cholesky(at::Tensor const&, bool) + 47 (0x10840eaef in liblantern.dylib)
#> frame #10: _lantern_cholesky_tensor_bool + 78 (0x10840e46e in liblantern.dylib)
#> frame #11: cpp_torch_namespace_cholesky_self_Tensor(XPtrTorchTensor, XPtrTorchbool) + 56 (0x1069fba38 in torchpkg.so)
#> frame #12: _torch_cpp_torch_namespace_cholesky_self_Tensor + 283 (0x10667fa5b in torchpkg.so)
#> frame #13: R_doDotCall + 837 (0x101ae0b05 in libR.dylib)
#> frame #14: bcEval_loop + 99716 (0x101b40c04 in libR.dylib)
#> frame #15: bcEval + 628 (0x101b1a134 in libR.dylib)
#> frame #16: Rf_eval + 506 (0x101b1987a in libR.dylib)
#> frame #17: R_execClosure + 761 (0x101b1c3b9 in libR.dylib)
#> frame #18: applyClosure_core + 128 (0x101b1b4c0 in libR.dylib)
#> frame #19: Rf_eval + 1189 (0x101b19b25 in libR.dylib)
#> frame #20: do_docall + 599 (0x101aa8857 in libR.dylib)
#> frame #21: bcEval_loop + 34028 (0x101b30b6c in libR.dylib)
#> frame #22: bcEval + 628 (0x101b1a134 in libR.dylib)
#> frame #23: Rf_eval + 506 (0x101b1987a in libR.dylib)
#> frame #24: R_execClosure + 761 (0x101b1c3b9 in libR.dylib)
#> frame #25: applyClosure_core + 128 (0x101b1b4c0 in libR.dylib)
#> frame #26: Rf_eval + 1189 (0x101b19b25 in libR.dylib)
#> frame #27: do_set + 2673 (0x101b20631 in libR.dylib)
#> frame #28: Rf_eval + 990 (0x101b19a5e in libR.dylib)
#> frame #29: do_begin + 429 (0x101b1eced in libR.dylib)
#> frame #30: Rf_eval + 990 (0x101b19a5e in libR.dylib)
#> frame #31: Rf_eval + 990 (0x101b19a5e in libR.dylib)
#> frame #32: do_eval + 1253 (0x101b20da5 in libR.dylib)
#> frame #33: bcEval_loop + 34028 (0x101b30b6c in libR.dylib)
#> frame #34: bcEval + 628 (0x101b1a134 in libR.dylib)
#> frame #35: Rf_eval + 506 (0x101b1987a in libR.dylib)
#> frame #36: forcePromise + 230 (0x101b1a366 in libR.dylib)
#> frame #37: Rf_eval + 634 (0x101b198fa in libR.dylib)
#> frame #38: do_withVisible + 57 (0x101b21139 in libR.dylib)
#> frame #39: do_internal + 362 (0x101b8c0aa in libR.dylib)
#> frame #40: bcEval_loop + 34655 (0x101b30ddf in libR.dylib)
#> frame #41: bcEval + 628 (0x101b1a134 in libR.dylib)
#> frame #42: Rf_eval + 506 (0x101b1987a in libR.dylib)
#> frame #43: forcePromise + 230 (0x101b1a366 in libR.dylib)
#> frame #44: Rf_eval + 634 (0x101b198fa in libR.dylib)
#> frame #45: forcePromise + 230 (0x101b1a366 in libR.dylib)
#> frame #46: getvar + 253 (0x101b431fd in libR.dylib)
#> frame #47: bcEval_loop + 18591 (0x101b2cf1f in libR.dylib)
#> frame #48: bcEval + 628 (0x101b1a134 in libR.dylib)
#> frame #49: Rf_eval + 506 (0x101b1987a in libR.dylib)
#> frame #50: R_execClosure + 761 (0x101b1c3b9 in libR.dylib)
#> frame #51: applyClosure_core + 128 (0x101b1b4c0 in libR.dylib)
#> frame #52: Rf_eval + 1189 (0x101b19b25 in libR.dylib)
#> frame #53: do_eval + 1253 (0x101b20da5 in libR.dylib)
#> frame #54: bcEval_loop + 34028 (0x101b30b6c in libR.dylib)
#> frame #55: bcEval + 628 (0x101b1a134 in libR.dylib)
#> frame #56: Rf_eval + 506 (0x101b1987a in libR.dylib)
#> frame #57: R_execClosure + 761 (0x101b1c3b9 in libR.dylib)
#> frame #58: applyClosure_core + 128 (0x101b1b4c0 in libR.dylib)
#> frame #59: R_forceAndCall + 535 (0x101b1ba17 in libR.dylib)
#> frame #60: call_loop + 311 (0x1018e0877 in purrr.so)
#> frame #61: map_impl + 253 (0x1018e0afd in purrr.so)
#> frame #62: R_doDotCall + 13245 (0x101ae3b7d in libR.dylib)