Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -2.2052  0.9325  4.9191 -1.7827 -5.3177
#>  -1.7477  2.8895  4.9752  0.8642 -5.1142
#>   4.0612 -0.2180 -3.7882  0.4444  2.0403
#>   0.5387  7.3962 -2.8093  7.9180 -3.4283
#>   4.7873 -1.0982 -7.2072  4.2582 -4.6686
#> 
#> (1,2,.,.) = 
#>    1.9106  -1.9547   0.0113   5.6538   0.6221
#>   11.5731  -1.5665  -3.1320  -7.6273   0.2431
#>   -1.2312  -3.7505  -1.1284  -8.7889  -3.6529
#>    5.8265 -14.1316  11.6525  -9.5437   4.6507
#>    1.1965  -3.8957  -3.6194  11.2987  -5.8419
#> 
#> (1,3,.,.) = 
#>    2.1773   1.8312   1.9625  -6.4510  -3.1584
#>    2.7755  -9.2318   7.5877   1.9592  -1.2194
#>    2.5895  10.1237   8.9430  -3.8181  -0.1899
#>   -9.5262  -5.2523  -2.5444   6.4063   7.6302
#>    5.7476   8.3121   2.4414  -3.7762  -2.6789
#> 
#> (1,4,.,.) = 
#>    1.8581  -3.4194  -0.1545   1.8134   1.5368
#>    9.3604  -0.6605   0.0998  -6.7827  -0.3819
#>    0.5821 -12.4539  11.8852 -10.7478  -2.2779
#>    3.4724 -22.0427  -2.7836  -4.1030  -2.5003
#>    7.8186   3.4934  -0.8856   3.3048  -0.2457
#> 
#> (1,5,.,.) = 
#>    0.5464   4.0138   0.0960  -9.0920   2.3494
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]