Div

Div

Arguments

input
(Tensor) the input tensor.

other
(Number) the number to be divided to each element of input

div(input, other, out=None) -> Tensor

Divides each element of the input input with the scalar other and returns a new resulting tensor.

Each element of the tensor input is divided by each element of the tensor other. The resulting tensor is returned.

$$ \text{out}_i = \frac{\text{input}_i}{\text{other}_i} $$ The shapes of input and other must be broadcastable . If the torch_dtype of input and other differ, the torch_dtype of the result tensor is determined following rules described in the type promotion documentation . If out is specified, the result must be castable to the torch_dtype of the specified output tensor. Integral division by zero leads to undefined behavior.

Warning

Integer division using div is deprecated, and in a future release div will perform true division like torch_true_divide. Use torch_floor_divide (// in Python) to perform integer division, instead.

$$ \text{out}_i = \frac{\text{input}_i}{\text{other}} $$ If the torch_dtype of input and other differ, the torch_dtype of the result tensor is determined following rules described in the type promotion documentation . If out is specified, the result must be castable to the torch_dtype of the specified output tensor. Integral division by zero leads to undefined behavior.

Examples

a = torch_randn(c(5))
a

#> torch_tensor 
#> -0.4509
#> -1.4741
#>  1.4410
#> -1.7545
#>  0.7006
#> [ CPUFloatType{5} ]

torch_div(a, 0.5)

#> torch_tensor 
#> -0.9018
#> -2.9482
#>  2.8819
#> -3.5089
#>  1.4013
#> [ CPUFloatType{5} ]

a = torch_randn(c(4, 4))
a

#> torch_tensor 
#>  2.0242  0.0625  1.1603  1.3928
#>  0.4491  0.0361  0.4277  1.9981
#> -0.5807 -0.4782 -0.2684 -1.0164
#> -0.6413  0.6253 -0.9699 -0.3891
#> [ CPUFloatType{4,4} ]

b = torch_randn(c(4))
b

#> torch_tensor 
#> -0.9792
#> -0.4541
#>  3.7302
#> -1.7271
#> [ CPUFloatType{4} ]

torch_div(a, b)

#> torch_tensor 
#> -2.0671 -0.1377  0.3111 -0.8064
#> -0.4586 -0.0795  0.1147 -1.1570
#>  0.5930  1.0529 -0.0720  0.5885
#>  0.6549 -1.3770 -0.2600  0.2253
#> [ CPUFloatType{4,4} ]