Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   2.7111  2.0122
#>   3.9755 -2.5458
#>  -1.3322 -2.9936
#> 
#> (2,.,.) = 
#>  -0.0873  2.5338
#>  -1.1463 -1.9630
#>  -3.3206 -9.1149
#> 
#> (3,.,.) = 
#>  -3.6823 -0.2882
#>  -1.4973  5.6480
#>  -0.6076  5.1892
#> 
#> (4,.,.) = 
#>   1.3374  4.6850
#>   3.8788  0.1292
#>   2.7394  0.5908
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>   0.0697  2.2357
#>   1.3027  0.3171
#>  -0.6303 -1.5821
#> 
#> (2,.,.) = 
#>   2.1362  0.2189
#>   1.8913 -3.3047
#>   2.2453 -3.5607
#> 
#> (3,.,.) = 
#>  -0.5553 -1.3737
#>  -0.0636  1.2340
#>  -0.3396  2.6799
#> 
#> (4,.,.) = 
#>   1.0605  0.9313
#>   0.8452 -0.7922
#>  -2.6076 -0.5307
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   8.7363  3.8818
#>   4.0321 -11.8082
#>   8.1930  1.4006
#>  -5.1731 -5.7308
#>  -3.9132  6.1517
#> 
#> (2,1,1,.,.) = 
#>  -4.6585 -4.0666
#>  -2.3974 -3.0817
#>   1.6630  7.1672
#>  -15.5327 -1.3414
#>  -4.8558  4.9128
#> 
#> (3,1,1,.,.) = 
#>   2.4472 -0.5163
#>  -4.5027 -2.4484
#>  -0.7576 -1.7906
#>   2.1669 -2.3029
#>  -0.7108  7.4406
#> 
#> (1,2,1,.,.) = 
#>   2.1113  0.4313
#>   5.3353 -1.0466
#>   1.4832  6.2818
#>   1.4419 -5.4126
#>  -3.5207 -8.0327
#> 
#> (2,2,1,.,.) = 
#>  -8.9937 -0.1250
#>   0.3340 -7.1360
#>   4.6932 -1.8468
#>   1.4689 -2.4392
#>  -1.5182 -9.2184
#> 
#> (3,2,1,.,.) = 
#>  -8.8077 -0.7060
#>   5.4882  8.0686
#>  -1.7359 -2.2111
#>  -3.0008  3.4189
#>   1.1613  1.6439
#> 
#> (1,3,1,.,.) = 
#>   1.4207 -2.3470
#>  -1.7842  4.7422
#>  -1.4133  2.5867
#>   3.7112  0.6948
#>   8.2447 -2.9379
#> 
#> (2,3,1,.,.) = 
#>   2.2821 -5.8034
#>  -3.4829  1.5684
#>   4.3098 -1.3261
#>  -3.7205 -0.4131
#>  -7.3670  1.4025
#> 
#> (3,3,1,.,.) = 
#>   0.8602 -2.3308
#>  -4.3080  1.1598
#>   3.2652  1.8645
#>  -4.6782 -7.4151
#>  -6.2663  4.3185
#> 
#> (1,1,2,.,.) = 
#>  -4.0530  3.8391
#>  -7.3503 -1.8232
#>   3.3822  2.1880
#>  -8.7822  9.4402
#>  -0.4459  0.3186
#> 
#> (2,1,2,.,.) = 
#>   9.4270  0.3410
#>   7.9194 -0.4462
#>   1.5181  5.1243
#>   3.7679 -4.9534
#>  -6.9161 -2.4061
#> 
#> (3,1,2,.,.) = 
#>   0.1855 -1.8936
#>  -5.8897 -7.0276
#>   6.0791 -2.2720
#>   1.9292 -5.0116
#>  -12.3559 -2.5689
#> 
#> (1,2,2,.,.) = 
#>   9.4659  0.0487
#>   3.6221 -2.8506
#>  -0.8181  7.0811
#>  -3.1900 -5.6312
#>   4.5978  3.3082
#> 
#> (2,2,2,.,.) = 
#>  -3.3408 -0.5115
#>  -10.2488 -3.1753
#>  -7.3446  2.2269
#>   2.1464 -0.1452
#>  -3.9005 -3.1766
#> 
#> (3,2,2,.,.) = 
#>   1.9773  1.7648
#>   5.1150 -3.1832
#>  -4.2779 -1.3763
#>  -3.6394  3.6146
#>   2.0338  5.3203
#> 
#> (1,3,2,.,.) = 
#>   1.6423  3.3517
#>   2.4120 -2.9111
#>   5.9619  6.1815
#>   1.6580  5.1172
#>   4.9902  4.3634
#> 
#> (2,3,2,.,.) = 
#>   1.5091 -3.9880
#>  -2.5021 -3.7183
#>   0.6893 -0.9951
#>  -5.8483  7.4076
#>  -7.4990  5.1157
#> 
#> (3,3,2,.,.) = 
#>  -5.7294  0.7694
#>  -0.9628 -0.5967
#>   2.1556  1.3675
#>  -2.6143  3.8830
#>   2.9878 -6.3074
#> 
#> (1,1,3,.,.) = 
#>   2.1507 -3.9811
#>   5.4291 -2.5809
#>  -3.1371 -3.2992
#>  -7.5630 -14.7583
#>   5.6223 -4.1573
#> 
#> (2,1,3,.,.) = 
#>   5.6866  2.7522
#>   2.8654  3.5473
#>  -6.4790  2.0200
#>  -2.9625 -2.7896
#>   0.7216 -7.2538
#> 
#> (3,1,3,.,.) = 
#>    6.2695  -1.5583
#>    3.9328  -1.1018
#>   12.5542   0.2857
#>  -10.0234  -3.9784
#>   -3.0514  -0.9776
#> 
#> (1,2,3,.,.) = 
#>   4.6380  4.2753
#>  -4.2851  1.9538
#>   7.8482 -2.5362
#>   0.4602 -4.3823
#>  -1.9308 -1.8711
#> 
#> (2,2,3,.,.) = 
#>  -0.7161  1.9117
#>   3.8956  7.4664
#>   3.8208 -4.0869
#>   3.6212  1.2198
#>  -7.2024  0.8249
#> 
#> (3,2,3,.,.) = 
#>  -0.4801  2.2609
#>   0.7042  2.0114
#>   3.3103 -3.9094
#>  -3.4338  5.3836
#>   2.8601 -10.7151
#> 
#> (1,3,3,.,.) = 
#>    2.6161   3.1170
#>    0.0944  -1.9508
#>   -2.9257  -3.2691
#>   14.0132   3.3209
#>    0.8359   2.2975
#> 
#> (2,3,3,.,.) = 
#>    3.7505  -3.6371
#>   12.8722   4.3380
#>    4.2726   1.0224
#>   -1.6101  11.5566
#>   -3.8600   2.6009
#> 
#> (3,3,3,.,.) = 
#>   -2.3385   6.7761
#>    1.5917  -4.7114
#>    1.7824   5.2551
#>    8.8486  -3.9957
#>   12.5274   6.1277
#> 
#> (1,1,4,.,.) = 
#>  -1.5351  1.3072
#>   3.0742 -3.1996
#>  -4.8258  1.6180
#>  -2.4355 -5.8734
#>   1.2933 -4.4533
#> 
#> (2,1,4,.,.) = 
#>   2.7579  0.5142
#>   4.4398  0.9486
#>  -5.1806 -2.3229
#>   5.7386  0.8960
#>   5.4199 -3.4645
#> 
#> (3,1,4,.,.) = 
#>  -1.0143 -0.4012
#>   1.6208  4.7890
#>  -0.0555 -3.9271
#>  -5.7271  0.4425
#>  -5.1920 -4.8471
#> 
#> (1,2,4,.,.) = 
#>  -7.1314  3.7633
#>   2.5568 -6.2649
#>   3.7672  9.0797
#>  -1.0928 -2.5722
#>   3.3110  5.4980
#> 
#> (2,2,4,.,.) = 
#>  -0.6124  0.2650
#>  -0.0963 -2.5194
#>  -1.4332  5.1896
#>   1.0619 -4.7393
#>  -0.9886 -4.8768
#> 
#> (3,2,4,.,.) = 
#>   0.6494 -0.1817
#>  -12.0875 -1.1678
#>  -5.5137 -0.4388
#>  -2.8013  0.3071
#>  -7.5697 -4.3397
#> 
#> (1,3,4,.,.) = 
#>  -1.9563 -4.9759
#>  -2.1901  0.4610
#>  -4.6464 -0.9092
#>  -6.2585 -7.6318
#>   3.0186 -5.6640
#> 
#> (2,3,4,.,.) = 
#>  -1.5244  6.8958
#>  -3.2556 -4.2445
#>   1.2582 -5.0919
#>   6.6932  9.7519
#>   2.8882 -0.8172
#> 
#> (3,3,4,.,.) = 
#>   -2.7865 -15.7157
#>    0.1499  13.3328
#>    3.3535   1.6580
#>    1.4621   0.6745
#>    1.3182   0.2735
#> 
#> (1,1,5,.,.) = 
#>   2.3069 -6.7114
#>  -11.8140 -3.0670
#>   3.4839  1.3903
#>   4.9735 -0.9836
#>   3.9707  4.3636
#> 
#> (2,1,5,.,.) = 
#>  -5.8182  0.5079
#>   1.2355 -2.3551
#>  -7.1970  6.0609
#>  -10.0189  4.1395
#>   3.7499  5.3732
#> 
#> (3,1,5,.,.) = 
#>  -0.2782  1.7240
#>  -7.1651 -0.3036
#>   2.2903  0.8543
#>  -4.3399 -2.3447
#>  -11.2251 -0.5272
#> 
#> (1,2,5,.,.) = 
#>   -5.2771   1.5134
#>   10.4245  -8.7358
#>    2.8709   2.2017
#>  -10.4751   1.0489
#>   -1.5564   1.0329
#> 
#> (2,2,5,.,.) = 
#>   2.3673 -10.5899
#>  -1.0902  8.2342
#>   3.2286 -2.6159
#>  -2.3542  7.2527
#>   6.0746 -7.1204
#> 
#> (3,2,5,.,.) = 
#>  -7.7486 -6.3777
#>  -0.0499  8.6834
#>  -11.1927  5.9029
#>  -5.3238 -1.6736
#>  -0.8603 -5.4320
#> 
#> (1,3,5,.,.) = 
#>  -2.8882  4.5331
#>   0.2974 -3.1758
#>  -0.6687 -1.2479
#>   7.4645 -0.3102
#>   1.3997  0.4341
#> 
#> (2,3,5,.,.) = 
#>  -1.9913 -11.1172
#>  -8.6550 -3.4361
#>   1.8042  2.3690
#>   5.0973  6.2363
#>   9.4120  0.3240
#> 
#> (3,3,5,.,.) = 
#>    3.1232  -3.0325
#>    5.2731  -2.1917
#>   11.8182   4.6997
#>   -5.5097  -3.7013
#>    0.0850  -2.3821
#> [ CPUFloatType{3,3,5,5,2} ]