Le

Le

Arguments

input
(Tensor) the tensor to compare

other
(Tensor or float) the tensor or value to compare

out
(Tensor, optional) the output tensor that must be a BoolTensor

le(input, other, out=None) -> Tensor

Computes \(\mbox{input} \leq \mbox{other}\) element-wise.

The second argument can be a number or a tensor whose shape is broadcastable with the first argument.

Examples

# \dontrun{

torch_le(torch_tensor(matrix(1:4, ncol = 2, byrow=TRUE)),
         torch_tensor(matrix(c(1,1,4,4), ncol = 2, byrow=TRUE)))

#> torch_tensor 
#>  1  0
#>  1  1
#> [ CPUBoolType{2,2} ]

# }