Median

Median

Arguments

input
(Tensor) the input tensor.

dim
(int) the dimension to reduce.

keepdim
(bool) whether the output tensor has dim retained or not.

out
(tuple, optional) the result tuple of two output tensors (max, max_indices)

median(input) -> Tensor

Returns the median value of all elements in the input tensor.

median(input, dim=-1, keepdim=False, out=None) -> (Tensor, LongTensor)

Returns a namedtuple (values, indices) where values is the median value of each row of the input tensor in the given dimension dim. And indices is the index location of each median value found.

By default, dim is the last dimension of the input tensor.

If keepdim is True, the output tensors are of the same size as input except in the dimension dim where they are of size 1. Otherwise, dim is squeezed (see torch_squeeze), resulting in the outputs tensor having 1 fewer dimension than input.

Examples

a = torch_randn(c(1, 3))
a

#> torch_tensor 
#>  1.5362 -1.4032  0.6089
#> [ CPUFloatType{1,3} ]

torch_median(a)

#> torch_tensor 
#> 0.608883
#> [ CPUFloatType{} ]

a = torch_randn(c(4, 5))
a

#> torch_tensor 
#>  0.8665 -0.2775 -0.5280 -0.9086 -0.5305
#> -0.0766  0.7030  0.2714 -1.9572  0.1124
#> -0.2088 -0.0836 -0.0411 -2.1958 -1.5354
#> -1.2656 -0.2068  0.0691 -1.3210  0.0746
#> [ CPUFloatType{4,5} ]

torch_median(a, 1)

#> [[1]]
#> torch_tensor 
#> -0.5280
#>  0.1124
#> -0.2088
#> -0.2068
#> [ CPUFloatType{4} ]
#> 
#> [[2]]
#> torch_tensor 
#>  2
#>  4
#>  0
#>  1
#> [ CPULongType{4} ]
#>