Min

Min

Arguments

input
(Tensor) the input tensor.

dim
(int) the dimension to reduce.

keepdim
(bool) whether the output tensor has dim retained or not.

out
(tuple, optional) the tuple of two output tensors (min, min_indices)

other
(Tensor) the second input tensor

Note

When the shapes do not match, the shape of the returned output tensor follows the broadcasting rules .

min(input) -> Tensor

Returns the minimum value of all elements in the input tensor.

min(input, dim, keepdim=False, out=None) -> (Tensor, LongTensor)

Returns a namedtuple (values, indices) where values is the minimum value of each row of the input tensor in the given dimension dim. And indices is the index location of each minimum value found (argmin).

Warning

indices does not necessarily contain the first occurrence of each minimal value found, unless it is unique. The exact implementation details are device-specific. Do not expect the same result when run on CPU and GPU in general.

If keepdim is True, the output tensors are of the same size as input except in the dimension dim where they are of size 1. Otherwise, dim is squeezed (see torch_squeeze), resulting in the output tensors having 1 fewer dimension than input.

min(input, other, out=None) -> Tensor

Each element of the tensor input is compared with the corresponding element of the tensor other and an element-wise minimum is taken. The resulting tensor is returned.

The shapes of input and other don't need to match, but they must be broadcastable .

$$ \text{out}_i = \min(\text{tensor}_i, \text{other}_i) $$

Examples

a = torch_randn(c(1, 3))
a

#> torch_tensor 
#>  1.8920 -0.4077 -1.2080
#> [ CPUFloatType{1,3} ]

torch_min(a)

#> torch_tensor 
#> -1.20805
#> [ CPUFloatType{} ]

a = torch_randn(c(4, 4))
a

#> torch_tensor 
#> -0.1830 -0.9674 -0.2161  0.8314
#> -0.2002 -0.0094 -1.0684 -1.4885
#>  0.7536 -1.6282  1.5211 -2.1920
#> -1.6796  1.0721 -1.7980  1.3457
#> [ CPUFloatType{4,4} ]

torch_min(a, dim = 1)

#> [[1]]
#> torch_tensor 
#> -0.9674
#> -1.4885
#> -2.1920
#> -1.7980
#> [ CPUFloatType{4} ]
#> 
#> [[2]]
#> torch_tensor 
#>  1
#>  3
#>  3
#>  2
#> [ CPULongType{4} ]
#> 

a = torch_randn(c(4))
a

#> torch_tensor 
#>  0.2454
#> -1.7728
#> -0.0927
#>  0.2862
#> [ CPUFloatType{4} ]

b = torch_randn(c(4))
b

#> torch_tensor 
#> -0.8052
#> -1.1960
#> -0.3272
#> -0.6723
#> [ CPUFloatType{4} ]

torch_min(a, other = b)

#> torch_tensor 
#> -0.8052
#> -1.7728
#> -0.3272
#> -0.6723
#> [ CPUFloatType{4} ]