Mul

Mul

Arguments

input
NA

value
(Number) the number to be multiplied to each element of input

out
NA

input
(Tensor) the first multiplicand tensor

other
(Tensor) the second multiplicand tensor

out
(Tensor, optional) the output tensor.

mul(input, other, out=None)

Multiplies each element of the input input with the scalar other and returns a new resulting tensor.

$$ \text{out}_i = \text{other} \times \text{input}_i $$ If input is of type FloatTensor or DoubleTensor, other should be a real number, otherwise it should be an integer

Each element of the tensor input is multiplied by the corresponding element of the Tensor other. The resulting tensor is returned.

The shapes of input and other must be broadcastable .

$$ \text{out}_i = \text{input}_i \times \text{other}_i $$

Examples

a = torch_randn(c(3))
a

#> torch_tensor 
#> -1.3128
#>  0.0966
#> -0.4937
#> [ CPUFloatType{3} ]

torch_mul(a, 100)

#> torch_tensor 
#> -131.2837
#>    9.6596
#>  -49.3712
#> [ CPUFloatType{3} ]

a = torch_randn(c(4, 1))
a

#> torch_tensor 
#>  0.6866
#> -1.0891
#> -1.8984
#> -2.3084
#> [ CPUFloatType{4,1} ]

b = torch_randn(c(1, 4))
b

#> torch_tensor 
#>  1.2499  0.5791 -1.5868  1.4754
#> [ CPUFloatType{1,4} ]

torch_mul(a, b)

#> torch_tensor 
#>  0.8582  0.3976 -1.0895  1.0130
#> -1.3612 -0.6307  1.7281 -1.6068
#> -2.3727 -1.0994  3.0123 -2.8008
#> -2.8852 -1.3369  3.6630 -3.4058
#> [ CPUFloatType{4,4} ]