Normal

Normal

Arguments

mean
(Tensor) the tensor of per-element means

std
(Tensor) the tensor of per-element standard deviations

generator
(torch.Generator, optional) a pseudorandom number generator for sampling

out
(Tensor, optional) the output tensor.

size
(int...) a sequence of integers defining the shape of the output tensor.

Note

When the shapes do not match, the shape of mean is used as the shape for the returned output tensor

normal(mean, std, *, generator=None, out=None) -> Tensor

Returns a tensor of random numbers drawn from separate normal distributions whose mean and standard deviation are given.

The mean is a tensor with the mean of each output element's normal distribution

The std is a tensor with the standard deviation of each output element's normal distribution

The shapes of mean and std don't need to match, but the total number of elements in each tensor need to be the same.

normal(mean=0.0, std, out=None) -> Tensor

Similar to the function above, but the means are shared among all drawn elements.

normal(mean, std=1.0, out=None) -> Tensor

Similar to the function above, but the standard-deviations are shared among all drawn elements.

normal(mean, std, size, *, out=None) -> Tensor

Similar to the function above, but the means and standard deviations are shared among all drawn elements. The resulting tensor has size given by size.

Examples

if (FALSE) {
torch_normal(mean=0, std=torch_arange(1, 0, -0.1))

torch_normal(mean=0.5, std=torch_arange(1., 6.))

torch_normal(mean=torch_arange(1., 6.))

torch_normal(2, 3, size=list(1, 4))
}