Randint_like

Randint_like

Arguments

input
(Tensor) the size of input will determine size of the output tensor.

low
(int, optional) Lowest integer to be drawn from the distribution. Default: 0.

high
(int) One above the highest integer to be drawn from the distribution.

dtype
(torch.dtype, optional) the desired data type of returned Tensor. Default: if None, defaults to the dtype of input.

layout
(torch.layout, optional) the desired layout of returned tensor. Default: if None, defaults to the layout of input.

device
(torch.device, optional) the desired device of returned tensor. Default: if None, defaults to the device of input.

requires_grad
(bool, optional) If autograd should record operations on the returned tensor. Default: False.

memory_format
(torch.memory_format, optional) the desired memory format of returned Tensor. Default: torch_preserve_format.

randint_like(input, low=0, high, dtype=None, layout=torch.strided, device=None, requires_grad=False,

memory_format=torch.preserve_format) -> Tensor

Returns a tensor with the same shape as Tensor input filled with random integers generated uniformly between low (inclusive) and high (exclusive).

.. note: With the global dtype default (torch_float32), this function returns a tensor with dtype torch_int64.

Examples