Rfft

Rfft

Arguments

input
(Tensor) the input tensor of at least signal_ndim dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

onesided
(bool, optional) controls whether to return half of results to avoid redundancy. Default: True

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

rfft(input, signal_ndim, normalized=False, onesided=True) -> Tensor

Real-to-complex Discrete Fourier Transform

This method computes the real-to-complex discrete Fourier transform. It is mathematically equivalent with ~torch.fft with differences only in formats of the input and output.

This method supports 1D, 2D and 3D real-to-complex transforms, indicated by signal_ndim. input must be a tensor with at least signal_ndim dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary, where \(N_i\) is the size of signal dimension \(i\).

The real-to-complex Fourier transform results follow conjugate symmetry:

$$ X[\omega_1, \dots, \omega_d] = X^*[N_1 - \omega_1, \dots, N_d - \omega_d], $$ where the index arithmetic is computed modulus the size of the corresponding dimension, \(\ ^*\) is the conjugate operator, and \(d\) = signal_ndim. onesided flag controls whether to avoid redundancy in the output results. If set to True (default), the output will not be full complex result of shape \((*, 2)\), where \(*\) is the shape of input, but instead the last dimension will be halfed as of size \(\lfloor \frac{N_d}{2} \rfloor + 1\).

The inverse of this function is ~torch.irfft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

x = torch_randn(c(5, 5))
torch_rfft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   2.5211  0.0000
#>  -0.9325  4.3145
#>   2.2256  6.6970
#> 
#> (2,.,.) = 
#>   1.7562  6.0036
#>   7.3650 -1.9397
#>  -5.9835  0.8725
#> 
#> (3,.,.) = 
#>   2.6427 -2.3389
#>  -2.5532 -1.8579
#>   0.6688  3.8394
#> 
#> (4,.,.) = 
#>   2.6427  2.3389
#>  -1.5704 -7.7015
#>   0.3130 -2.9953
#> 
#> (5,.,.) = 
#>   1.7562 -6.0036
#>   5.1782 -0.0964
#>  -2.6110  3.4039
#> [ CPUFloatType{5,3,2} ]

torch_rfft(x, 2, onesided=FALSE)

#> torch_tensor 
#> (1,.,.) = 
#>   2.5211  0.0000
#>  -0.9325  4.3145
#>   2.2256  6.6970
#>   2.2256 -6.6970
#>  -0.9325 -4.3145
#> 
#> (2,.,.) = 
#>   1.7562  6.0036
#>   7.3650 -1.9397
#>  -5.9835  0.8725
#>  -2.6110 -3.4039
#>   5.1782  0.0964
#> 
#> (3,.,.) = 
#>   2.6427 -2.3389
#>  -2.5532 -1.8579
#>   0.6688  3.8394
#>   0.3130  2.9953
#>  -1.5704  7.7015
#> 
#> (4,.,.) = 
#>   2.6427  2.3389
#>  -1.5704 -7.7015
#>   0.3130 -2.9953
#>   0.6688 -3.8394
#>  -2.5532  1.8579
#> 
#> (5,.,.) = 
#>   1.7562 -6.0036
#>   5.1782 -0.0964
#>  -2.6110  3.4039
#>  -5.9835 -0.8725
#>   7.3650  1.9397
#> [ CPUFloatType{5,5,2} ]