Std_mean

Std_mean

Arguments

input
(Tensor) the input tensor.

unbiased
(bool) whether to use the unbiased estimation or not

dim
(int or tuple of ints) the dimension or dimensions to reduce.

keepdim
(bool) whether the output tensor has dim retained or not.

std_mean(input, unbiased=True) -> (Tensor, Tensor)

Returns the standard-deviation and mean of all elements in the input tensor.

If unbiased is False, then the standard-deviation will be calculated via the biased estimator. Otherwise, Bessel's correction will be used.

std_mean(input, dim, unbiased=True, keepdim=False) -> (Tensor, Tensor)

Returns the standard-deviation and mean of each row of the input tensor in the dimension dim. If dim is a list of dimensions, reduce over all of them.

If keepdim is True, the output tensor is of the same size as input except in the dimension(s) dim where it is of size 1. Otherwise, dim is squeezed (see torch_squeeze), resulting in the output tensor having 1 (or len(dim)) fewer dimension(s).

If unbiased is False, then the standard-deviation will be calculated via the biased estimator. Otherwise, Bessel's correction will be used.

Examples

a = torch_randn(c(1, 3))
a

#> torch_tensor 
#>  1.9597 -0.4032 -1.0707
#> [ CPUFloatType{1,3} ]

torch_std_mean(a)

#> [[1]]
#> torch_tensor 
#> 1.59225
#> [ CPUFloatType{} ]
#> 
#> [[2]]
#> torch_tensor 
#> 0.161936
#> [ CPUFloatType{} ]
#> 

a = torch_randn(c(4, 4))
a

#> torch_tensor 
#> -0.5568  1.5005  0.4914  2.4290
#> -0.8748 -0.3086 -0.0217  0.4132
#>  1.5194 -0.1060 -0.9213  0.1436
#>  1.1747 -1.3945 -0.2690  0.0256
#> [ CPUFloatType{4,4} ]

torch_std_mean(a, 1)

#> [[1]]
#> torch_tensor 
#>  1.2872
#>  0.5400
#>  1.0146
#>  1.0557
#> [ CPUFloatType{4} ]
#> 
#> [[2]]
#> torch_tensor 
#>  0.9660
#> -0.1980
#>  0.1589
#> -0.1158
#> [ CPUFloatType{4} ]
#>