Transpose

Transpose

Arguments

input
(Tensor) the input tensor.

dim0
(int) the first dimension to be transposed

dim1
(int) the second dimension to be transposed

transpose(input, dim0, dim1) -> Tensor

Returns a tensor that is a transposed version of input. The given dimensions dim0 and dim1 are swapped.

The resulting out tensor shares it's underlying storage with the input tensor, so changing the content of one would change the content of the other.

Examples

x = torch_randn(c(2, 3))
x

#> torch_tensor 
#>  0.8275  0.2687  0.4284
#>  0.2057 -0.2063  0.0312
#> [ CPUFloatType{2,3} ]

torch_transpose(x, 0, 1)

#> torch_tensor 
#>  0.8275  0.2057
#>  0.2687 -0.2063
#>  0.4284  0.0312
#> [ CPUFloatType{3,2} ]