Context-manager that enable anomaly detection for the autograd engine.

This does two things:

Running the forward pass with detection enabled will allow the backward pass to print the traceback of the forward operation that created the failing backward function.

Any backward computation that generate "nan" value will raise an error.

with_detect_anomaly(code)

Arguments

code
Cod that will be execued in the detect anomaly context.

Warning

This mode should be enabled only for debugging as the different tests will slow down your program execution.

Examples

if (torch_is_installed()) {
x <- torch_randn(2, requires_grad = TRUE)
y <- torch_randn(1)
b <- (x^y)$sum()
y$add_(1)

try({

b$backward()

with_detect_anomaly({
  b$backward()
})

})

}

#> Error in (function (self, inputs, gradient, retain_graph, create_graph)  : 
#>   one of the variables needed for gradient computation has been modified by an inplace operation: [CPUFloatType [1]] is at version 1; expected version 0 instead. Hint: enable anomaly detection to find the operation that failed to compute its gradient, with torch.autograd.set_detect_anomaly(True).