Distribution is the abstract base class for probability distributions. Generates a sample_shape shaped sample or sample_shape shaped batch of samples if the distribution parameters are batched. Generates a sample_shape shaped reparameterized sample or sample_shape shaped batch of reparameterized samples if the distribution parameters are batched. In PyTorch: sample_shape=torch.Size() Returns the log of the probability density/mass function evaluated at value. Returns the cumulative density/mass function evaluated at value. Returns the inverse cumulative density/mass function evaluated at value. Returns tensor containing all values supported by a discrete distribution. The result will enumerate over dimension 0, so the shape of the result will be (cardinality,) + batch_shape + event_shape (where event_shape = ()for univariate distributions). Note that this enumerates over all batched tensors in lock-step[0, 0, 1, 1, ...]. With expand=FALSE, enumeration happens along dim 0, but with the remaining batch dimensions being singleton dimensions, [0, 1, ... @param expand (bool): whether to expand the support over the batch dims to match the distribution's batch_shape. @return Tensor iterating over dimension 0. Returns entropy of distribution, batched over batch_shape. @return Tensor of shape batch_shape. Returns perplexity of distribution, batched over batch_shape. @param Tensor of shape batch_shape. Returns the size of the sample returned by the distribution, given a sample_shape`. Note, that the batch and event shapes of a distribution instance are fixed at the time of construction. If this is empty, the returned shape is upcast to (1,).

Distribution is the abstract base class for probability distributions. Generates a sample_shape shaped sample or sample_shape shaped batch of samples if the distribution parameters are batched. Generates a sample_shape shaped reparameterized sample or sample_shape shaped batch of reparameterized samples if the distribution parameters are batched. In PyTorch: sample_shape=torch.Size() Returns the log of the probability density/mass function evaluated at value. Returns the cumulative density/mass function evaluated at value. Returns the inverse cumulative density/mass function evaluated at value. Returns tensor containing all values supported by a discrete distribution. The result will enumerate over dimension 0, so the shape of the result will be (cardinality,) + batch_shape + event_shape (where event_shape = ()for univariate distributions). Note that this enumerates over all batched tensors in lock-step[0, 0, 1, 1, ...]. With expand=FALSE, enumeration happens along dim 0, but with the remaining batch dimensions being singleton dimensions, [0, 1, ... @param expand (bool): whether to expand the support over the batch dims to match the distribution's batch_shape. @return Tensor iterating over dimension 0. Returns entropy of distribution, batched over batch_shape. @return Tensor of shape batch_shape. Returns perplexity of distribution, batched over batch_shape. @param Tensor of shape batch_shape. Returns the size of the sample returned by the distribution, given a sample_shape`. Note, that the batch and event shapes of a distribution instance are fixed at the time of construction. If this is empty, the returned shape is upcast to (1,).

Distribution is the abstract base class for probability distributions. Generates a sample_shape shaped sample or sample_shape shaped batch of samples if the distribution parameters are batched. Generates a sample_shape shaped reparameterized sample or sample_shape shaped batch of reparameterized samples if the distribution parameters are batched. In PyTorch: sample_shape=torch.Size() Returns the log of the probability density/mass function evaluated at value. Returns the cumulative density/mass function evaluated at value. Returns the inverse cumulative density/mass function evaluated at value. Returns tensor containing all values supported by a discrete distribution. The result will enumerate over dimension 0, so the shape of the result will be (cardinality,) + batch_shape + event_shape (where event_shape = ()for univariate distributions). Note that this enumerates over all batched tensors in lock-step[0, 0, 1, 1, ...]. With expand=FALSE, enumeration happens along dim 0, but with the remaining batch dimensions being singleton dimensions, [0, 1, ... @param expand (bool): whether to expand the support over the batch dims to match the distribution's batch_shape. @return Tensor iterating over dimension 0. Returns entropy of distribution, batched over batch_shape. @return Tensor of shape batch_shape. Returns perplexity of distribution, batched over batch_shape. @param Tensor of shape batch_shape. Returns the size of the sample returned by the distribution, given a sample_shape`. Note, that the batch and event shapes of a distribution instance are fixed at the time of construction. If this is empty, the returned shape is upcast to (1,).

Methods

Public methods

Distribution$new()

Distribution$expand()

Distribution$sample()

Distribution$rsample()

Distribution$log_prob()

Distribution$cdf()

Distribution$icdf()

Distribution$enumerate_support()

Distribution$entropy()

Distribution$perplexity()

Distribution$.extended_shape()

Distribution$.validate_sample()

Distribution$.get_checked_instance()

Distribution$print()

Distribution$clone()

Method new()

Usage

Distribution$new(batch_shape, event_shape, validate_args = NULL)

Method expand()

Usage

Distribution$expand(batch_shape, .instance = NULL)

Method sample()

Usage

Distribution$sample(sample_shape = NULL)

Method rsample()

Usage

Distribution$rsample(sample_shape = NULL)

Method log_prob()

Usage

Distribution$log_prob(value)

Method cdf()

Usage

Distribution$cdf(value)

Method icdf()

Usage

Distribution$icdf(value)

Method enumerate_support()

Usage

Distribution$enumerate_support(expand = TRUE)

Method entropy()

Usage

Distribution$entropy()

Method perplexity()

Usage

Distribution$perplexity()

Method .extended_shape()

Usage

Distribution$.extended_shape(sample_shape = NULL)

Arguments

sample_shape

(torch_Size): the size of the sample to be drawn. Argument validation for distribution methods such as log_prob, cdf and icdf. The rightmost dimensions of a value to be scored via these methods must agree with the distribution's batch and event shapes.

Method .validate_sample()

Usage

Distribution$.validate_sample(value)

Arguments

value

(Tensor): the tensor whose log probability is to be computed by the log_prob method. Returns a dist_constraint object representing this distribution's support. Returns the mean on of the distribution Returns the variance of the distribution Returns the standard deviation of the distribution

Method .get_checked_instance()

Usage

Distribution$.get_checked_instance(cls, .instance = NULL)

Method print()

Usage

Distribution$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

Distribution$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.