Computes the Cholesky decomposition of a complex Hermitian or real symmetric positive-definite matrix.

Letting be or , the Cholesky decomposition of a complex Hermitian or real symmetric positive-definite matrix is defined as

linalg_cholesky(A)

Arguments

A

(Tensor): tensor of shape (*, n, n) where * is zero or more batch dimensions consisting of symmetric or Hermitian positive-definite matrices.

Details

A=LLHL∈Kn×n A = LL^{H}\mathrlap{\qquad L \in \mathbb{K}^{n \times n}} A=LLHL∈Kn×n

where is a lower triangular matrix and is the conjugate transpose when is complex, and the transpose when is real-valued.

Supports input of float, double, cfloat and cdouble dtypes. Also supports batches of matrices, and if A is a batch of matrices then the output has the same batch dimensions.

See also

Examples

if (torch_is_installed()) {
a <- torch_eye(10)
linalg_cholesky(a)
}
#> torch_tensor
#>  1  0  0  0  0  0  0  0  0  0
#>  0  1  0  0  0  0  0  0  0  0
#>  0  0  1  0  0  0  0  0  0  0
#>  0  0  0  1  0  0  0  0  0  0
#>  0  0  0  0  1  0  0  0  0  0
#>  0  0  0  0  0  1  0  0  0  0
#>  0  0  0  0  0  0  1  0  0  0
#>  0  0  0  0  0  0  0  1  0  0
#>  0  0  0  0  0  0  0  0  1  0
#>  0  0  0  0  0  0  0  0  0  1
#> [ CPUFloatType{10,10} ]