Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.3582 -0.2159 -0.1827  0.8054  0.8876 -0.3912 -0.4746  1.4137  0.0386  0.0404
#>  0.3445 -1.0158 -0.9891 -1.4501  0.8460 -0.2724  1.5548 -0.2529  0.6746 -0.7987
#>  0.6423  0.8965 -1.6293 -0.2374  2.1833 -0.2664  0.6866  0.0454 -1.2862  0.7574
#>  1.3154 -0.0276 -0.3454 -0.6552  0.0536  1.3640 -0.0378 -1.3019  0.4443 -0.7111
#>  0.1661  0.6525  0.5288 -1.0878 -0.2896 -0.2208 -0.5060 -0.9796 -0.3413 -0.8606
#> -1.8120  0.5255 -1.4915 -0.2189 -0.6770 -0.3698  0.5289  1.2556 -1.6123  0.4226
#>  1.6313 -0.0682 -0.6068 -0.2849  0.5327 -0.4219 -0.2274  0.5274 -0.5134  1.2904
#>  0.0579  1.9324  0.4883  0.5633  1.2850 -2.4354 -1.9457  0.7371  0.2991  0.1437
#>  0.3693 -1.1321  0.1913  0.2601 -0.0921  0.6448  0.1840  0.1711 -0.3450 -0.4708
#>  1.6985 -0.9381  1.3864 -0.1050 -1.1469  0.1657  0.2951  0.9964  0.6064  0.0455
#>  0.1021  0.0726  1.2640  0.4398  2.4957 -0.5626 -0.9199  0.5785  0.1431 -0.6428
#>  1.7992  1.4287 -0.4438  1.9916  0.0858  0.7700  0.4455  0.1178 -0.0103  0.1917
#> -0.9927 -1.4311  0.3323  0.4792 -0.9850 -0.8706  1.3476  0.0392  0.4365  0.0978
#>  0.0634  0.1048  0.6288 -0.8110 -0.4781  1.8002  1.6457  0.7376  0.5218 -0.0765
#>  0.3951 -0.3380 -0.1846 -0.1421  1.4849 -0.8163  0.8512  0.0535  0.4632  2.2465
#>  0.2540  1.0485 -0.2560 -0.2887  0.9630  0.0318 -0.0100  1.2395  0.8678  0.4887
#> -0.6720  1.4646  0.1034  0.6737  0.9649  1.2214 -0.0263  0.6284  0.1731 -0.7567
#> -0.1268  1.3355  0.0347 -0.1992  0.4429  1.2409 -0.5310 -1.3388 -0.0861 -0.6841
#>  0.6591  0.8917 -0.4785 -0.3710  2.0013 -0.5296 -1.2669  0.5473  0.4757 -0.0731
#> -0.0834 -0.3886  0.7776  0.2722  1.0847 -0.1191  0.3201 -0.5152  0.0992  0.2572
#>  0.7104 -1.0037  0.7780 -0.7753 -1.1805 -0.3091  0.9033 -0.2192 -1.0872  0.5308
#>  0.7430  0.6646  0.1113  0.0283  0.0381  0.6447  0.1118  0.9692  0.4765  0.5470
#>  0.5214  0.3729  1.9009 -0.6511  0.3467  0.1885 -1.5508  0.0834 -0.8375 -0.6994
#> -0.7528  0.2889 -1.3356 -0.2317  0.0506 -0.4236  2.1370  0.7324  0.7938 -1.1187
#>  0.4978 -1.1632 -0.2012  0.3300 -0.2863  0.9262 -0.9998  0.8491  0.8303 -0.0798
#>  0.0323  0.6173  0.1495  0.2212 -0.3531 -1.5332  1.0286 -0.5044  2.0726 -0.2344
#> -1.6815 -0.8169 -0.2648 -0.5408 -0.4165 -0.6875  2.4275 -0.4371  0.8844 -0.3520
#>  0.2448  0.7383 -0.7875  0.2774  1.4197 -0.2094  0.8172  1.4010 -0.9935 -0.0317
#>  1.0500  0.9287  0.1556  0.6041 -2.1321  0.3018  0.9262 -1.1854  1.7083 -0.1758
#>  0.7569 -0.1197 -1.3990 -1.3412  0.1393  0.7262 -0.9606  0.4541  2.1033  0.1798
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]