Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.8984  0.2295 -0.4453 -0.1059  0.4648 -2.2994  0.5075  2.3224 -2.1466 -0.2474
#> -0.0037 -0.0617  0.1369  1.9373 -0.9462  1.0075  1.3810  1.0860  0.5338  0.7630
#> -1.8403 -0.6340 -0.4708 -0.0752 -0.4471  0.7731  0.3219 -0.7945  0.3517 -2.4127
#> -0.8024  0.5147  0.2700 -0.9077 -0.7503 -0.1428 -0.0886  0.3914  0.2979 -1.2210
#> -1.2840  0.4378  0.7697  1.6506 -0.6497 -1.4002  2.0860  0.9451  0.9716  0.9847
#>  0.0668  0.7604  0.4456  0.2452  0.0480  1.3698  0.7787  0.3744 -0.0803  1.1316
#>  0.1916  0.8953 -0.8031  0.4300  0.4754  0.2318 -1.7410  1.7486 -1.6046  0.3664
#>  0.3021  0.3183  1.7547  0.7255  0.8122 -1.0636 -0.6385 -1.7048 -0.2831  0.0355
#> -1.0799  0.9165  2.1627  0.3544 -1.0085 -1.4786  1.1303 -0.7129  0.3831  0.9320
#> -0.6760 -0.5296  1.2071 -1.5258  1.3158 -0.3532 -0.7512  1.9196 -0.4799 -2.4748
#>  0.4275  0.6532 -0.7604 -0.2286 -0.0415 -1.1634 -0.5987 -0.9342  1.6185  1.4842
#> -0.9934 -0.3063 -2.4496 -1.2419 -0.0016 -0.9208 -0.0816 -1.1729  0.6587  0.4963
#>  1.0174  0.4708  1.1470  1.6168  0.4937 -0.2144 -1.2530 -1.0354  1.4368 -1.2634
#> -1.6712  1.5547 -0.6471 -0.8592 -0.8158 -1.5268 -0.5384  0.0165  0.5777 -0.1055
#> -0.0655  0.0812 -2.4348 -0.0845  0.7797 -1.7756 -0.2909  0.0735  0.7268  0.1429
#>  0.4641  0.5279  0.8943  0.0292 -2.2035 -1.6959 -0.2920 -0.0565  1.5362 -0.1619
#> -0.8083 -0.9419  0.2167  0.8747 -1.9275 -1.8039 -2.5575  0.9865 -0.8185  0.4912
#>  0.3511 -0.2108  0.1685 -1.4079 -0.0596 -1.7532  0.6340 -1.1960  2.9025  0.2579
#>  0.2476 -0.9424 -0.1610 -2.1467  0.7237  0.0434 -0.2589 -0.4727 -0.4740 -0.3544
#> -0.9596  1.4062  2.6648 -0.3265  3.1580 -0.1326 -0.3760 -0.4019 -1.4004 -0.3786
#>  0.9420  0.1714 -1.2475 -2.0350 -0.2228  0.7410  1.2664 -0.1204  0.4267  0.7120
#>  0.0421 -0.8373 -0.9665 -0.3952 -0.1753  1.0078  1.7535  0.4857  0.1596 -1.5556
#>  1.7638  0.5787  0.3658 -1.7900  1.4119  0.1112  0.0841 -0.6522  0.9963  0.0938
#>  1.3354  0.2436 -0.7918  0.6403  0.1960  1.0502  1.2457  2.4221 -0.7917  0.5630
#> -0.1949 -0.5427 -0.9313 -1.0185  0.9738  0.0022 -0.8175 -1.2568 -1.1349  0.5358
#>  0.8193  1.2423 -0.5376 -2.4263  0.9417 -0.2138 -0.0231 -0.8544  0.2757  0.3017
#>  0.8271 -1.9703 -0.3729 -0.0123 -0.6083  0.5947  0.0018 -0.1456  1.1050  0.3043
#> -1.6148  0.1541  0.3203 -0.9067  0.0291  0.4149 -0.4319 -0.7738 -0.1942 -0.4506
#>  0.9918 -2.0406  1.3174  0.1026 -0.8120  1.6174 -0.6625  1.4564  1.6986 -0.3320
#>  1.2015 -0.5893  1.2546  0.1714 -0.0600  0.7027  0.0566 -0.2818 -0.6396  0.5690
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]