Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.1475  0.4172 -0.2598  0.2789  0.3518  0.8777 -0.8303  0.5163  0.4808  0.8378
#> -0.3070  0.3473  0.9483  0.8456 -0.3559  0.5910 -1.1436  0.4323  1.1654  1.4169
#>  1.2848 -0.1821  1.3509  0.1495  0.0191  0.6076 -1.6076  0.9465 -1.0805 -1.1369
#>  0.3611  1.3616 -0.1922 -0.1200 -0.3009  0.4410 -1.1201  0.2032  1.3258 -0.8433
#>  0.0316  1.9725  0.5414 -0.8272 -1.3135  0.7799  0.3354  0.2004  0.0616  0.6876
#>  0.2145  1.2068 -0.6813  0.4766  0.2644  0.3325 -0.6944  0.2560  1.5704 -0.1296
#> -0.7228 -0.5313  0.6298 -1.5389  0.5907 -1.4322 -0.1266  1.8740  0.3764  0.0781
#> -2.2456  0.2754 -0.7858 -0.0058  2.0229 -0.1272 -1.1267  0.5926 -1.0089 -1.2767
#> -1.3748  0.0846 -0.3036  0.3613  1.2127 -0.3517 -0.0053 -0.4245 -0.3916  1.1726
#> -1.0498 -1.4956 -0.3862  1.2987 -0.8373  0.1159  0.1852 -0.6960  1.0383 -0.2417
#> -0.3649 -1.0491 -0.0981 -0.4184 -0.0941 -0.2098  2.0074  0.2160  0.5696 -1.3921
#> -0.5488  1.2959 -1.2728 -0.6949 -1.4561 -0.5089 -1.3563 -0.9262  0.3265 -0.2791
#> -1.7549  0.3902  1.0742 -1.2047  0.6435 -0.3239 -0.1942 -1.4980  0.3888  1.1400
#> -0.6924  0.1487 -0.4190  0.5382  1.3414  0.1053 -3.4292 -1.6962  1.0618  0.0160
#> -1.0259 -0.2738  1.1829 -0.7259 -0.6172  0.3303 -0.3494  0.8454 -0.8520 -0.3165
#>  0.5066 -0.1552 -0.6295 -0.8220  0.0315 -2.8376 -0.2381  0.4380  0.9279  2.0809
#>  0.3283  0.1634 -1.2973  0.3240  0.0088  1.5074  0.7464  1.2209  1.4003 -0.2055
#> -0.1838  1.4749  0.0873 -0.5616  0.8917 -0.8470 -1.1999 -1.6993  0.0179  0.2647
#>  1.3377 -0.7566  3.5316  1.3662 -0.2585 -0.3806 -0.1443 -1.0673 -0.1206 -0.4892
#> -0.3099  1.2561 -0.4640 -0.1085  1.2975  1.6432  0.0140  1.5059 -1.1052 -1.1234
#> -0.1908  1.5414  0.4513  0.7850  1.3994  1.0987 -0.0360  0.1156  1.2204  0.4154
#>  0.9588 -1.1670 -1.5348  0.9013  0.7047  0.0921 -0.2011 -0.9847  0.5416 -0.3319
#>  1.0899 -1.5851 -0.7591  1.9046 -2.7920  0.4872 -2.3532 -1.4247  1.0874 -1.3308
#>  0.4194  0.6207 -1.1393  0.5581 -0.4559  2.1403 -0.1615 -0.0768  0.5605  0.1339
#> -0.0847 -0.7169  1.5679 -0.2848 -0.7652  2.1707 -0.5181  1.4043 -0.1609  1.4337
#> -1.1424  0.9804 -1.0008  0.5145 -0.0816  2.0860 -0.7766  0.6824  0.8333 -0.9018
#>  0.7049 -1.0696  0.7055  0.3558 -0.0696 -1.4897  1.4534 -0.8854  1.1453 -0.0035
#>  0.6113 -1.6817 -0.0055 -0.9302  0.3524  0.2667  0.3667 -0.0895 -0.9015 -1.0445
#>  0.0571 -0.7978  0.5940 -0.6414 -1.6102 -0.4705  1.5236 -0.2606 -0.7896  2.5968
#> -0.4973  1.3102  1.1768  1.5712  0.8775  0.8287 -0.2083  0.3537  0.9048  0.0775
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]