Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.7511 -0.3836  0.3891  0.3571  0.6343 -0.4977 -0.1639 -1.7609 -0.6844  1.6135
#> -0.0873  0.8079 -1.2552 -2.1382 -0.1954 -1.4849  1.3529  1.5733  0.4835  0.7089
#> -0.3796  0.0295 -1.2949  0.0022 -0.8002  0.9829 -0.0474  0.5638 -0.7700  2.0921
#>  1.2205  1.2803 -0.8990  1.1712 -0.5847 -1.8779 -1.4134  0.8845  1.3362 -0.2474
#> -1.3315 -0.3835  0.9071  1.2505  0.7975  2.0614  1.6140 -0.5392  1.3665 -2.9047
#>  0.6009 -1.9642 -0.1671 -1.8974 -0.2840  0.3927  0.6339 -0.4679 -0.1042 -1.6293
#> -1.7345 -0.6391  0.2120  0.8440  0.1767  0.0578 -1.3345 -0.1248 -0.0934 -0.8466
#>  0.0406  0.7903 -0.5498 -0.1081  0.3722  0.3977 -1.3411 -0.0053  0.4288  0.7700
#>  1.1175 -0.6078 -0.3749  1.1549 -0.5404 -1.5896 -0.0730 -0.7107 -0.4427  0.8762
#> -1.3735 -2.8798 -1.2712 -0.9716 -0.9243  0.6383  0.1752  0.1458  0.7742  1.1110
#> -0.8333  0.5501  0.4248  1.0976 -0.9264  0.7364  0.2539 -0.1364  0.8163 -0.3302
#> -0.0321 -1.0826 -1.6056 -1.7044  0.9907  0.9494  0.5542  0.1356  0.6147  2.1849
#> -0.5863 -0.2080 -0.0300 -1.6706  0.7502  1.6020  1.2251  0.4416 -0.5424 -0.5347
#> -0.0168 -1.5815 -2.3415 -2.3221 -0.2197  0.0047 -0.9947 -0.4803  0.6148  1.3446
#> -0.2797  0.0529 -2.2556  2.5486 -0.0356 -0.6517 -0.7537  0.4939  0.1143  1.4658
#>  0.6672 -0.8119  2.0123  0.3688  0.0414 -1.0397 -0.2076 -0.1223 -0.9373 -0.5786
#> -0.1525  0.8298  0.8293  0.6362  0.6336  0.4514 -0.1391 -0.3253  0.1023 -1.9877
#> -0.6710 -1.1212  0.9559 -0.6959  0.1204  1.9232  0.0585 -0.2114  0.7166 -0.6513
#> -1.6137 -1.0324  0.7959 -0.5528  0.1758  0.3921  1.2485  1.1711  1.0758  0.9106
#> -0.6958  0.6629 -1.2386  0.7607 -0.5680 -0.2636 -0.3086 -0.5110  0.4117 -0.3621
#>  0.2841  0.4134 -1.6240  0.8458 -0.4504  0.0987  0.8335  0.5002 -0.6841 -1.3150
#> -0.1451  0.4676 -0.4871  1.4795 -0.1823  0.4182 -1.2631 -0.3627 -0.6597 -0.3980
#>  1.9381  1.0719 -0.0205  0.2655  0.9532 -1.1819  0.0360 -2.1554 -0.0559 -1.2434
#>  0.2880 -0.6250 -0.2590 -0.5887 -0.6023 -1.4187 -0.3204  1.1983 -1.1931  0.2178
#> -2.0113  0.4696  0.6575 -0.0106  1.0613  1.0112  0.0615 -0.1904 -0.6229  0.6825
#> -0.7320  0.3599 -1.1548 -0.4650 -1.0823 -0.4295  0.0940  0.6662  0.2413  0.3751
#>  0.0159 -0.4117  0.0590  0.7857 -0.3707  0.5037 -0.0810 -0.0926  0.3498 -0.6456
#> -1.3612  0.0865 -1.2136 -0.5361  0.2903  0.1002 -2.3757 -0.2581 -0.3223  0.7107
#>  0.1066 -2.2629 -1.1789 -1.4495  0.4901 -0.0473 -1.1797 -0.6467  1.8663 -1.2005
#> -1.3308 -0.3017 -0.0545 -0.1490  0.9407 -2.6891 -1.2320 -1.6461 -2.0527 -2.0002
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]