Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.5965 -0.2175 -0.5631  0.0586 -1.4960  0.2805 -0.3160  0.0165 -0.5814 -0.3879
#>  0.0175 -1.4264 -0.5377  1.6456  0.5966  0.3933  2.1568 -1.3645 -0.3352 -0.8391
#>  0.2975 -2.1785 -0.1652  0.4322  0.4310 -1.0473  2.5921 -0.6576  0.0443 -1.4703
#>  0.5922  0.5920  0.7677 -0.3368  0.6362  0.6798  0.1884  2.4414 -0.4481  0.4694
#> -0.3001  0.4711 -1.8560  0.8170  0.2278  1.9114  1.8067  0.5399 -0.5047  0.0629
#>  0.7066 -1.4115 -1.1983 -0.2123  0.3989  1.3940  2.0188  0.6302 -1.4299 -1.7889
#> -1.5439 -0.9109  1.7485 -1.2726 -1.5363  1.4230 -0.0857 -0.9252  0.8773  1.2619
#>  1.1740  0.3059  0.1557 -0.2138  0.1387  0.0329 -0.4751 -1.1783  0.8317 -0.6032
#> -0.0998 -1.3186  0.2055 -0.4646 -0.5070 -0.1971 -1.6562  1.7421  0.1767  1.0934
#> -1.4503  0.8810  0.7110  0.7979 -0.4664 -0.0381 -0.5988 -2.1856  0.9563  0.2912
#> -0.0718 -1.4873  1.2527  0.5360 -1.1684 -0.1756 -0.1188  0.2859  0.8118 -0.7416
#>  0.2798 -1.0038 -0.7658 -0.3852 -1.1685  0.7211 -1.0265  0.7420  1.2643  0.1385
#>  0.3321  0.1894  0.2525  0.9807  1.3095 -0.5577 -1.1114  1.2346  0.7485 -0.6439
#> -1.0750 -0.7745  0.1703 -1.3276  0.5276 -0.4417 -0.6650  1.4441  1.3153 -0.2044
#> -2.2523 -0.8875 -0.2694  1.7207  0.8143 -0.1249 -0.6111  1.3319  0.8155 -0.2076
#>  0.1396  2.2848  0.2705 -1.4926 -1.8050 -0.9263 -1.5502 -0.5831 -0.1503 -0.4524
#> -0.6370  0.5212  0.4429  0.0453 -0.6730  0.6939 -1.1286 -1.1872  0.3501  1.3754
#>  0.6555  1.0811  0.5527 -0.9175 -0.1165 -0.2519 -1.7307  0.3552  0.3136 -0.4012
#> -1.2091 -1.1252 -0.2376  0.3922  0.6726  1.0036 -1.1316  1.4344  0.2129  1.5682
#> -0.0453 -0.3891  0.5937  0.0636 -0.8446  0.5992 -0.5838 -0.0144 -0.2798  1.0139
#> -0.3843  1.4136 -1.7169  1.6067 -1.3920  0.2745 -0.2622 -0.6291  3.0460 -0.8625
#>  1.5721  1.4227  0.6312 -1.4738  0.6865  0.4007  0.9940  2.2446  1.6089  1.6859
#> -0.1596  0.2858  0.5381  1.4686  0.5466 -0.8092 -0.0442  0.0065  0.0214  1.0166
#>  0.6879  0.2422  1.4730  1.5206 -1.1946 -1.1995  0.8214  0.4096  0.4001 -1.2640
#>  1.4643 -0.8683 -2.4384  0.9689  0.7722  0.3759 -1.7225 -0.7584  2.5878  0.3185
#>  1.3302 -0.3506 -1.7283 -0.4776  0.9119  0.0154 -0.0148  0.4250  1.3267  0.5904
#> -1.1581  0.3182 -0.3760  0.0097  0.1568 -0.7460 -1.0693  0.5838  1.1608 -1.5324
#> -0.8069  2.2866 -0.7020 -0.2336 -1.1183  0.7324 -0.5764  0.0586  0.5121 -1.6437
#>  0.0272  0.3050 -1.2688 -0.1008  0.4843 -1.0439 -0.0600  0.2057  0.2389  0.1237
#>  0.0663 -2.2310  0.1563  0.3041 -0.5925 -1.8326 -0.9960 -2.0675 -0.8043 -1.6030
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]