Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.5183 -0.0750  1.5305  0.4221 -1.4956  1.1423 -1.0510  1.3128  0.9665 -2.4187
#>  1.0801 -0.5092  0.6903 -0.3376 -1.1234  0.2292 -0.0342 -1.6609 -2.3539 -1.4330
#>  1.8214 -0.7075  1.2192  1.6059 -1.2986 -0.8659  0.9396 -1.1007 -1.5946 -0.4701
#> -1.8131  3.5477  0.5753 -1.2325 -0.2189 -0.0763 -0.6297  0.3591  1.2661 -1.1805
#>  3.3544  0.3376 -2.9201 -0.4085 -0.3535  1.0728  1.9970 -0.5942  0.0351 -1.9970
#> -1.7289  0.4960  0.4112 -0.7341  0.7510  0.3948 -0.4885 -0.6427 -0.6423  0.7853
#>  0.9542  0.9073  0.3647 -1.1592  1.2376 -0.5767 -1.2322 -0.2364 -0.9907  0.1433
#>  0.4886  0.1845  1.1193  0.4625  0.5882 -0.0981  1.2703  1.7203  0.9474 -1.2716
#> -0.4541  1.9928  0.7023 -2.2044  0.7141 -0.4595 -0.2424 -0.5833 -0.9141  0.4725
#> -0.4647  1.0847 -0.8453  0.7112 -0.0492  0.7879  0.4296 -0.1446 -1.9403 -1.3499
#>  0.0018  2.7795  1.2498  0.4924 -0.3330 -0.4539 -1.3984  0.5008  0.2688 -0.3543
#> -1.1531 -0.3138  1.0108  2.1431 -1.1190 -1.2092 -0.4514  0.4302 -0.4716 -2.2672
#> -2.0846  0.2183  1.9547  0.3597 -1.0258 -0.8951 -0.2018 -0.3478  0.2803  1.3163
#> -0.1541  0.8292 -1.4005 -0.3349  0.1696  1.8788 -0.1687 -0.1380 -0.2799 -0.5948
#> -0.5049 -0.5165 -0.6379 -0.6559 -1.3819 -0.1531 -1.6228  0.0935  0.4218 -0.4697
#> -0.1198 -0.2249 -0.5256 -1.6251 -0.9148  0.6704 -0.5093 -0.8717  0.0213 -0.2426
#> -0.0718 -0.4034 -0.6733  0.0461 -0.9577 -0.7039 -1.5858 -2.0164  0.0541 -0.7672
#> -0.2986  0.1468 -0.4635  0.2763  0.8074  0.2762 -0.3115 -0.4440 -0.2414 -0.2703
#>  0.7470 -0.2944 -0.7326 -1.2494 -0.6544  0.7804 -1.7742  0.9261  1.2714 -1.5532
#>  0.5764 -1.0069 -0.2016  3.3647  1.6182  1.1397  0.5314  0.5754 -0.7334 -0.6695
#> -0.9995 -1.1378 -1.3169  1.4869 -0.0772 -0.0319  0.5507  0.5009  1.2212 -1.0880
#>  0.0046  2.4037 -0.2208  0.8327 -1.2024 -0.2920  0.2813  0.1316  0.5175 -0.1580
#> -0.1577 -1.2348 -1.8530  0.7308  0.7044 -0.6554  0.4693  0.4869 -1.4041 -0.2530
#> -0.4718  2.3398  1.0569  0.6269  0.7256 -0.2643  0.6669 -0.4668  1.0583 -0.9269
#> -0.8745  0.9799 -0.0287  0.6394  0.8394  0.2777  0.3199 -0.8906  0.7876  0.4385
#> -0.2101  1.5596 -0.6911 -0.2572  0.2028 -0.5380 -0.4442  1.4747 -0.5042  1.2590
#>  0.9789  1.4045 -2.5844 -0.2443 -0.2103 -0.1745 -0.5705 -0.6312 -0.2845 -2.2457
#>  1.2980 -0.4288  0.8832 -1.2506 -1.1128  1.6336 -0.3990 -1.2050 -0.4673 -2.0853
#> -0.8479 -1.0451 -0.1069 -0.3670  0.8901 -0.4571  0.0854 -0.9202 -0.5856  1.0983
#>  1.5969 -0.0340 -0.3786 -2.5758 -0.4081  0.4031  1.2828  0.3821 -0.2421 -0.0599
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]