Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.9576 -0.2774  1.7085 -1.1166  1.6888 -0.1497 -0.7050  0.2207  0.6130  1.1355
#> -1.2796  0.9946  1.1680 -0.5807  1.2956 -1.0248  0.0241  0.9716 -0.6625  0.3391
#>  1.7853 -1.3681  1.6640  0.1592  1.2989  0.4886  1.5710 -2.4735 -1.0932  0.1006
#>  0.5440 -0.7608  0.4824 -0.4473  1.3399  2.4670  0.0061  1.3109 -1.2269 -0.1901
#> -1.9093 -0.8101 -0.4835 -0.0015 -1.3624  0.9046 -0.9340  0.2887  0.4974  0.2610
#> -1.6894  0.9391 -0.0397  0.2613 -3.1028 -0.9805  0.6144 -0.2886 -0.2125  0.8022
#> -0.6704 -0.2887  1.1381 -1.7288  0.8380 -0.2090 -1.2318  1.0561  1.0298 -0.9846
#>  0.5375  1.9583 -0.8985 -1.5838 -1.1853 -0.0172 -0.6405 -1.4772  0.3930  0.3453
#> -1.9195  1.1476  0.3136 -0.1246 -0.8953  1.1859  0.6130 -0.3231 -0.5700  0.1752
#>  1.2564  0.4113 -1.3256  0.4314  1.4689  1.6228  1.3350 -0.5900  1.0234 -0.8159
#>  1.1529  0.4536  0.6869  1.6511 -1.0136 -0.2506  1.0580  0.1613 -0.6985  0.9767
#> -0.7971 -0.2457  2.6876  0.8432  0.1077  0.8435  0.1738 -1.6362 -1.1093  1.4310
#> -0.0044  0.9029  0.2125 -0.2435  2.1330 -0.2372 -0.8943 -0.8261 -0.4386  0.2736
#> -0.7903  2.0631  0.8962  0.0349 -0.1243  0.1251 -0.7894  2.8731 -1.2986  0.3111
#> -3.1514  1.1119 -0.6832  0.6188 -0.1222  1.1755 -1.0967 -0.6440 -0.2582 -0.8287
#>  0.2356  1.6111  0.2143 -0.8114  0.6786 -0.3939  1.1701 -0.9650  0.8998  1.6919
#>  0.0797 -0.1220 -0.0316  0.0935  0.0955 -0.4723 -0.6881 -0.7985 -0.6906 -0.5346
#> -0.6554  0.5543 -0.2221 -0.1264  0.3566 -0.5028 -1.5993  0.9374 -0.8731  0.9947
#>  1.1367 -0.5384 -0.8891  2.1117  1.2881 -0.6474 -0.4713  0.0266 -0.8396  1.6327
#>  0.0537 -0.0683  0.8677  0.0088 -1.8329 -0.1399  0.6418 -0.8848 -0.4420  0.4398
#> -1.0225 -0.6570  0.3792 -0.2813  0.4048  0.0748 -0.2279  0.0716  0.2420 -1.6870
#>  0.4434  0.9124 -0.3639  4.0424 -2.7516 -0.0167  0.6711  0.4304 -0.9479  0.3113
#>  1.1937  1.1665 -0.9135  1.2254 -0.2394 -1.3692  2.0754 -0.6037 -1.7147 -1.7834
#>  0.1453 -0.3180 -0.3903  1.2288  0.6046 -0.2172 -0.2662 -0.4914  1.1777  0.8756
#> -0.5778 -0.2341  0.3165 -0.0037 -0.9968 -1.4036  0.3144  0.3994 -0.5038 -0.0970
#> -0.9970 -2.1212 -0.7188 -0.0967  0.8612 -1.4479 -0.9622 -0.0497  1.1984  0.0347
#>  0.3706  1.5959  0.3249  0.3484 -0.3046 -0.5489  1.2905 -0.3000 -0.2929  0.5766
#>  1.9917  0.0037  0.7018 -0.1244  0.2580  0.1636  1.9646 -0.4353 -1.1074 -0.6276
#>  0.7152  0.4994 -1.2727 -2.0188  0.0481 -0.8545 -0.3824 -0.3616  0.7892  0.0013
#> -1.0563 -0.3146  1.3317  0.0693 -0.1698 -1.6698  2.0235 -0.4032  0.4078  0.8971
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]