Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.8998 -1.1559 -0.3163 -1.9563 -0.2087  0.8119  0.0342  0.0431 -0.4286 -1.0450
#>  0.4740  0.1506 -0.8598 -0.2912  1.0495 -1.2722 -0.3997 -0.8433 -0.5027 -1.7376
#>  0.6190 -0.7018 -0.5915  0.4698  1.1436  1.3909 -0.3673  0.1209  0.5632  0.3500
#> -0.4264  2.0141 -1.5712 -1.2035  0.7383 -1.0138 -1.1455  1.4256 -1.2901  0.0587
#> -0.1903  0.0288 -0.3071 -1.0173 -0.7840 -0.8107  0.8571 -0.4313  0.2249 -0.5303
#> -0.4472  2.1885 -0.3150  0.0081  1.5369  0.9597  0.0835  1.0891 -0.6744  0.4047
#>  1.2343  1.4886 -0.7140  0.7365 -1.2245 -0.4259 -0.3033 -0.5003 -0.4969 -1.0341
#> -0.0848  1.0439  0.3019  0.0787  0.1064 -0.6549  0.4993  0.5998  0.6955 -1.1633
#> -0.8494  1.1397 -0.4590 -0.7449  0.9472 -0.3699 -0.0713  0.9035  0.1381 -0.5799
#>  1.1966  1.3831  1.7933 -0.3538  0.4082 -1.7677  0.5771 -0.3923 -1.1746  0.0868
#>  0.5121 -1.1536  0.5530  0.9828 -1.1081  0.0776 -1.9849 -1.6367  0.1344 -0.9491
#> -1.2204 -0.2298 -0.7586 -0.7716  1.5064 -0.3540  0.2783  0.7886  0.3290 -0.5360
#>  1.2843  0.0679 -0.6069  0.4056  0.0683  0.3539  0.3493  0.6758 -1.6508 -1.4006
#>  0.1539  0.3527 -0.2888  0.5198  1.6005  0.8694 -0.6665  1.0807 -0.8628 -1.1269
#>  0.1125  0.0883 -2.2841  0.6529  0.9028  0.1447 -0.8281  0.3344  0.1001 -2.2369
#>  0.8211  0.2317 -0.4544 -1.7459  1.0592 -0.4876 -0.8826  1.4271 -0.0530 -0.4390
#> -0.1629 -0.8200 -2.7333 -0.4788 -0.4471  0.7507 -0.0816 -0.8753  0.5071 -0.1832
#> -0.6921  0.2853 -0.3446  1.5857  1.6669 -0.0954  1.3267  0.9679 -0.5476 -0.3998
#>  1.8734 -0.2689 -0.4444 -0.5977 -0.1254  1.9060 -0.9127 -0.0960  1.2955 -1.0809
#> -0.2030 -0.6696 -1.1361  1.7809  0.1924 -0.1016 -0.9674 -1.1586 -1.5913  0.8689
#>  1.6554  0.7290 -0.3625  0.8932  0.0531  0.1356  2.0331  1.5604  0.7353  0.1450
#>  0.1857  0.7489 -0.1022  1.5901  0.8366  0.2374 -0.5116  0.5168 -1.2657 -1.5627
#>  1.8030 -2.0710 -0.6851  1.4709  0.8344  0.4690 -0.1762  1.1444  0.4215 -0.4155
#>  1.9839  0.0467 -1.6891 -0.6333 -0.2807 -0.2933  1.5214 -0.6359 -0.3298  0.4767
#> -0.0011  0.1744 -0.1230 -1.0432 -0.1043  1.6287  0.5001 -0.4881 -2.2232 -1.1619
#> -0.1489  0.1850  0.5869  0.5008  1.4706 -0.5783 -0.1647 -0.3134 -0.7901  0.7609
#> -0.5220  1.7743 -0.4935  1.0599 -0.1060 -0.6592 -0.0741 -0.2091  1.4165  0.9917
#> -1.5615 -0.1775 -0.8445  0.7598  0.1207 -2.8148 -0.1173 -1.0874  0.7308 -1.4926
#>  0.2946 -1.1676 -0.3502 -0.4089  0.3105 -0.9338 -1.0136 -0.1740 -0.5220  0.8631
#>  0.7418 -0.1822  0.1675  0.0047  0.6113  0.4924  0.1993  0.0339 -0.3547 -0.7373
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]