Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.2493  0.6997 -0.6034  0.1542 -0.0426 -0.2463  1.1238 -0.7960 -0.4550  0.4702
#>  1.5300  2.3225 -1.1619  0.7747  0.7216 -0.3405  0.5227 -0.5222  0.8005  0.2263
#> -0.6552  0.8978  0.5007 -0.4101  0.2072 -0.7385 -0.6887  0.0977  0.3105 -1.2341
#>  2.0511 -0.0875  1.0749 -0.5018 -0.3642 -0.3724  0.7009 -0.2421 -0.3949  1.1254
#>  0.1576 -0.5210  0.1229 -0.6257 -0.9808  1.5416 -0.8813 -0.1189 -0.0362  0.0488
#>  0.7791  0.2356  0.7712 -0.6159  2.0337  0.1224 -1.7120 -1.1344  0.5093 -0.3499
#> -0.0779 -1.2413 -0.8614 -2.2714  0.0931 -0.1960  1.1010  0.7715 -0.0267 -0.3014
#> -0.8418  0.3583 -1.1145 -0.5482  0.2991 -0.4732 -1.8403  0.9903  0.3692  2.1062
#>  0.9481 -0.0769 -0.7000 -0.2548 -2.0294  0.7792  1.2741  1.3234 -0.7789  1.2381
#> -1.0985  1.4838 -0.9851  0.5327  0.0639 -0.5011  0.6783  0.0556 -0.0135 -0.2162
#>  0.5261  0.2193 -0.8848 -0.3481 -1.2260  0.2196 -1.3302  0.2879  0.3790 -0.4544
#> -0.8555 -0.6046  0.6667 -2.0727  0.2223  0.1139  0.7752 -1.8992  0.7228 -0.4863
#> -0.3772 -0.7642 -0.5594  0.2661  1.6822  0.6612 -0.6996  0.9395 -1.3880  0.6378
#> -0.6900 -1.2218 -1.5668 -0.2714 -0.4865  0.5829  1.5120  0.2224 -0.4455 -0.4610
#>  1.5380  0.0229 -0.4780  0.5988  0.0992 -0.1844  0.5192 -1.2255 -0.6008 -2.1044
#>  0.6917  0.3007 -0.7027 -0.4127  0.2373 -1.0381  0.2695  0.1936  0.1725  0.3566
#>  0.4654  0.5707 -0.9601  0.5522  2.0812  0.6728 -0.5141  0.7527 -1.0123  0.3662
#> -0.8928 -0.6492 -0.4788  0.5134 -0.0936 -0.2850 -0.9188 -1.7539 -0.8160 -0.1366
#> -1.4519  0.0347 -0.5645  0.1465 -0.6151 -0.9386  1.2511  0.6014 -0.1693  0.7674
#> -1.5115 -1.8958  1.9466 -0.7470 -0.5055  0.4368  1.2825  0.5218 -1.0519  0.0782
#> -0.8356 -0.3652  1.1457 -0.8039  0.3009  0.1576  0.6253  0.2740  0.9713  0.9587
#> -0.6066 -0.0491  0.4630 -1.3767  0.2782 -0.4052 -0.5436  1.2289 -1.0503 -0.6641
#> -0.1569 -1.0326 -0.1911 -1.1796 -0.2121  0.3534  0.7542 -0.4680 -0.1704  0.9276
#> -0.8752 -0.4990 -1.8443  1.7813  0.9319 -1.1110  0.3014 -0.3697  0.2672  0.9672
#> -0.3584  1.9423 -0.5394 -0.4815  0.4345  0.3831  0.2851  0.3360  1.1199 -0.3719
#>  0.0596 -0.5281  0.1995  0.4400  0.1815 -0.3943  0.0537 -0.1600  0.6330  0.5499
#> -2.5630 -1.2246  1.4939 -0.4126 -0.2959 -0.6079 -1.3756 -1.4258 -0.9971 -0.4156
#>  0.3923  0.7818 -0.0872 -0.9175 -2.0083  0.3903 -0.8462 -0.4809 -0.6082 -1.5874
#>  1.4585 -0.4914  0.7808 -1.4766  0.2660  0.2305 -1.4440  1.1502  1.0623  0.5511
#> -1.1410  0.5202  0.7143  0.1955 -0.3034 -2.2584 -0.0256  1.0217 -1.4262 -1.0012
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]