Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.1695  0.3004 -0.4870  0.8296  1.7407  0.9463  0.3700  1.1995 -0.5837 -0.8782
#> -0.4623 -0.3314 -0.3059  0.8259 -0.3421  0.8471 -2.1876  0.7940  1.0863  0.8412
#>  0.3978 -1.6240 -1.0281  0.0763  0.8829  0.6244  0.4325 -0.3498 -1.6608  0.4029
#>  1.1233 -2.9208 -1.7936 -0.6106 -1.8123  0.0968  0.5472  0.3230  0.4409  0.9349
#> -0.4809 -0.0350  2.1345 -0.0174  0.2161 -1.7448 -0.2579  1.0233 -0.0793  0.0338
#> -0.3392 -0.0759  0.3507 -0.3916  0.6617 -0.4573  1.7192  0.4109 -0.2864 -0.8119
#>  0.2093 -1.1804  0.2199 -0.7466 -0.5215 -0.5237 -0.4453 -2.1330  0.9775  0.5053
#> -0.1138  0.7340 -0.0360 -0.8388 -0.2101  0.2799  0.6166 -0.1293 -0.5120 -0.3794
#>  0.5776  1.5050 -0.3135  0.0373  0.5405 -1.6103 -1.2011  0.3326  0.7613  1.0662
#>  0.2049  0.6193 -0.9455  1.5304 -0.5653 -0.2960  1.3360 -1.4229  0.0417  1.2928
#>  0.0601 -1.3300 -0.4928  0.1593 -1.4721 -0.0632 -0.0801  0.3223 -0.1326  0.3712
#>  0.6775 -0.0618  1.2040 -0.3133  0.1051  1.1867  1.5840 -0.8041  0.3653 -0.1866
#> -0.2252 -0.0603  0.0390  2.3757  0.2015 -0.7752 -0.8675  0.0204 -0.3920  0.0403
#>  0.2367  0.9120  1.7883  1.5046 -0.4573 -1.2416 -0.6295  0.4980 -1.3547 -0.8262
#> -0.6775 -0.6068  0.0442  0.5540 -0.2286 -0.6941 -1.6733  1.8899 -0.2945 -1.0391
#> -0.2053 -0.0489  1.0643 -0.8221  0.6556 -0.3264  0.3024  0.8516  0.0588  1.3166
#>  0.7725  0.3887 -1.2021  0.1011 -1.0919 -0.3491  0.9092  0.7927  0.3564  0.6853
#>  1.0045 -1.4058 -0.3157 -1.6497  0.9176 -2.3685  0.3131 -2.2524 -1.5989 -0.3896
#> -1.1031  0.2753 -0.2715 -0.5767 -0.2132  0.0321 -0.4992 -0.1429 -0.5356  0.1184
#> -1.9834 -2.0106 -0.5655 -1.2034  0.0823 -1.2758  0.3102  0.6856 -1.0684 -0.6673
#> -0.2834 -1.1387 -0.5190 -0.2570  1.5095 -0.8308 -0.1044 -1.8487  0.7553 -0.7313
#>  0.4138 -1.0240  0.2953 -0.4768  0.6673  2.0000 -0.4454 -0.3696 -0.3960  1.9734
#>  0.0903  0.6936  3.1640  2.4022 -0.5290 -0.3822  0.5839  0.1920 -1.7853  0.8155
#>  0.7148  0.4799 -0.1827  1.4600 -0.3613 -1.8703  0.4997 -0.7582 -0.8626 -1.0174
#>  1.9529 -1.0501  0.6992  0.5896  0.3393  2.5083  0.5182  1.6528  1.6715  0.9489
#>  0.0964  1.6939 -1.4603 -0.1185 -0.8675 -2.5680 -0.4343 -0.9472 -0.1844 -0.3739
#> -1.0997 -0.7073 -0.1520 -1.2441  1.9712 -1.8800 -0.5821 -0.3714 -0.7769 -0.8922
#> -1.8146  1.5090  0.6667 -1.0567 -0.4020 -2.3254  0.0551 -0.9660 -0.6019 -0.1786
#> -0.0162  1.3831  0.2334 -0.0519 -0.8342 -0.3800  0.1670  0.3616  0.4029  1.4375
#> -2.3743  0.3806  0.4182  0.9719  1.1323 -0.1768  0.6318 -0.4619  0.1979  1.7920
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]