Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-2.0668  1.4853  0.9092  0.7071  0.0511 -0.1022  1.1376  0.1323  0.7852 -0.5238
#>  0.4939  1.3750 -0.8483  1.1110 -1.2698  0.6504 -0.5443 -1.3265  0.1730 -1.4311
#>  0.0265  0.7609  1.3051  0.7853  0.2819  0.3256  0.2084 -0.6492 -2.5169  0.1133
#> -0.3026 -1.0283  0.1396  0.5983  0.0486 -0.2226  0.2769  0.8652 -1.0087  0.6670
#>  0.8497  0.4377  0.5790 -1.3811 -1.0748 -1.3225 -0.1073  2.2134  1.9861  1.1701
#>  0.0022 -2.7758  1.5573 -0.7553 -0.8806 -0.3035 -0.1000 -0.4066 -0.2689 -0.3863
#>  0.7528 -0.6463  0.5940 -0.1036  1.4599  0.2369 -0.0064  0.0596  1.1093 -1.2368
#> -1.5305  0.5800  0.4720 -0.5451 -0.3585  0.5226  0.0892 -0.5783 -0.9911 -0.1048
#>  0.3295 -2.3153 -0.1309  0.4841 -0.3405  1.2783  0.0597 -1.1307  1.5424 -0.3212
#> -0.3091  0.2587 -0.6750 -2.1162  0.3597  0.1490  0.0090  0.4266  0.3771  1.8413
#> -0.1347 -1.0406 -2.3394 -0.3489  3.4072 -0.5223  1.4898 -0.6077  0.0618 -0.7493
#>  2.2864 -0.3410  0.0041  0.2439 -0.3634  1.3611 -0.5371  0.5498 -1.0853  0.8351
#>  0.1067 -0.9445 -0.6663  1.0447 -0.6750 -0.3489  1.3968 -0.4733  1.1240  2.0384
#> -0.3722  0.6579  0.9668  0.9965 -0.0960 -0.8883  0.8883 -2.0981 -0.4432 -1.6336
#> -1.3220  0.3916  0.4851 -0.7334 -0.9387  1.1597  0.2025  1.3671  0.2548  1.3544
#>  0.0520 -0.6787  0.6322 -1.1152 -0.0705  0.3072  1.4135  1.1459  0.0534 -0.4961
#> -0.4182  0.3014 -1.3571 -0.1214 -2.0560  1.8753  0.1162  0.8080 -0.8558  1.3154
#>  1.8368 -1.4794 -0.7974  0.2276 -0.2628  0.9919 -0.3148  0.1451  1.4772  0.6913
#> -1.2038 -1.0297 -0.3816 -0.3140 -0.3995 -1.0411  0.3167  1.3070  1.0656 -0.8189
#>  0.9374  0.0413  0.0839  0.2809 -0.3043  2.2940  0.9339  0.0669 -0.7199  0.7202
#> -0.1966 -0.5148 -0.4285 -0.7683 -0.1695  0.9348 -2.8407 -1.2330  1.9924  0.9139
#> -0.2919 -0.5756  0.0367  2.2963  0.9679  1.4241  3.2055  0.7673  0.6728 -0.9848
#> -1.5661 -0.7310 -1.9150 -1.5644  1.2820 -0.5979 -0.9582  0.5453 -0.7733  0.7796
#> -1.0330 -1.1246  0.5762 -0.0495  0.9908  0.3537 -0.9928 -1.0020 -1.3129 -0.1085
#>  1.6097  0.4603  1.1803 -0.1746 -1.0929 -0.7203 -0.4075 -0.5375 -0.1615 -1.0287
#> -2.0735 -0.1325 -0.7999 -0.4360  1.2008  1.5258 -1.0396 -3.7657  1.2813 -0.4070
#> -1.8583 -0.2341  1.6613  0.9973  0.2246 -0.7163 -1.3725 -0.2247  0.3932 -0.6204
#> -0.8940 -0.6533  0.5823  0.3110 -1.5626  1.5900  0.7034 -0.1684 -0.0430 -1.0736
#>  1.7754 -0.9824 -0.8959  0.1151 -1.0538 -0.7804  1.4538 -1.1029 -0.5020 -1.3545
#> -0.2661  0.2276 -0.6380  1.0816  0.5969  2.5075  0.5845 -0.4374  0.2882 -1.2979
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]