Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 2.2762  1.3280  1.5804  1.2188 -0.3211 -0.4847  0.4014  0.6511  1.0446  0.3980
#> -0.1916  0.3563  0.3274  0.1267 -0.9498 -0.8701  0.6672  0.1520  1.2558  0.7844
#> -0.5132 -0.2003  1.1801 -0.6598 -1.8215 -1.1966 -1.3905 -0.8679  0.6151 -0.9349
#> -1.2200  1.3101 -0.1128 -0.3434  0.4816  0.0719 -1.6459  0.6979  1.2650  1.1117
#>  0.0901 -1.0726 -0.0611 -0.7058 -1.6561 -1.8739  0.5229 -0.3492  0.6639  0.6351
#>  1.8521 -0.2203 -0.9843  1.3806 -1.6840 -0.4411  2.6371 -0.8940 -0.6476  1.3280
#> -0.8484  0.5645  0.3850  0.8877  0.9011 -1.8224 -0.6593 -0.1354  0.6866 -0.7795
#>  2.1927 -0.3568 -0.3420 -2.6789 -0.7304  1.6694  0.2931 -0.3889 -1.6258 -0.9173
#> -0.9044 -1.5625 -0.9692 -0.8691 -0.7302 -0.0325 -0.4510 -1.4411  1.6038  0.7936
#> -0.1119  0.3400  0.4261 -0.2140  1.0493 -1.6433 -1.4136 -0.2980 -0.2524 -0.0091
#> -1.9512 -2.4043 -0.8892 -1.0203 -1.0694 -1.7929  0.6981  0.8191 -0.3291 -0.2303
#> -0.5235 -0.6219  1.4007 -0.8532  0.1222  0.1616  1.4699 -1.9719 -0.8447 -0.0082
#>  0.5824 -1.3462  0.4488 -0.8810  0.4181  1.5071  0.1596 -1.3742 -0.0580 -0.8461
#>  2.2131  0.0700 -0.8122  1.3601  0.5897 -0.5009  0.9404  0.4536 -0.0246 -1.0412
#>  0.2670 -0.3354  1.8194  0.6284  2.2392 -1.3139  0.5503  0.5342 -0.4839  0.7122
#> -0.0893 -2.2141 -1.8414  0.8006  1.3498  1.0121  0.9344  1.1557  0.2165 -1.9276
#>  0.0964  0.5889 -0.6210 -0.2134 -0.5229 -0.5038 -0.7134  1.0617 -3.1909  0.2496
#>  0.2656  0.7279  1.0895  0.0414  0.9405 -0.9179 -0.2642 -0.3020 -0.5943  1.3229
#> -0.7479  0.9672 -0.7855 -0.6833 -0.6825 -0.7873 -1.9516 -0.9990  0.1744 -0.4959
#> -0.0530  0.9755 -1.7451 -0.5325 -0.4032  0.0217  0.1898  0.9057 -1.7629 -1.0873
#>  1.3208 -0.7874 -2.2808 -0.8795  0.0439  0.5887  1.1837 -1.1488 -0.4216  0.0359
#> -1.0789  0.3961 -1.5894 -0.2304 -0.5855  0.8527  1.0799  0.9690 -0.9273  0.2887
#>  1.0103  0.3727 -0.0085 -0.0246  0.6676 -1.1527  0.6626 -1.8036 -0.5542 -0.5525
#> -1.3374 -0.7300 -1.3214  1.5111  2.2399 -1.1232  0.9700 -0.5740 -1.4228 -0.6350
#> -1.0241 -0.9779  0.2977  0.7832 -0.7287  0.9643 -0.4121  1.3559  1.8895  1.0026
#>  0.4525 -1.1903 -0.4376  0.0561 -0.4205 -0.6027  0.5851 -3.0528 -0.0008  1.7140
#> -0.1724 -1.4789  0.6185  0.8412 -0.8052  1.1435  1.3818  0.2495  0.4793  0.2921
#> -1.5968 -1.5981 -0.5125 -1.0704 -0.1530 -0.2333 -1.0658 -0.3043 -0.3205  0.0088
#> -1.4975 -0.6280  0.2690  1.4842 -0.6796  0.6260 -1.1956  0.2088  0.3642 -1.3370
#>  0.1803 -1.2655 -0.6954 -0.7396  0.2290 -0.7500  1.4618 -0.5884 -0.9093  0.4780
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]