Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.3733 -0.0093  0.0013  0.1225 -0.5975  0.5239 -0.3913  0.0682 -1.2147  0.8395
#> -0.7580  0.5202 -0.8866 -0.2763  0.0998  0.6682 -0.1854  0.4769 -0.7605  0.0963
#>  0.7182  0.1166  1.6407 -1.8811 -0.6574 -3.1979 -0.2277 -1.1821 -0.8045 -1.1719
#> -0.4107 -1.1651  0.3428 -1.4126  0.2589  1.3511  0.2530 -0.2695  0.2460  0.2768
#>  0.2122 -0.2201 -1.1759  1.0952 -0.7369  0.8286 -0.1036  1.6274  0.1421 -2.1897
#> -0.3084 -1.5018  0.2249  0.6798  1.1202  1.6254  1.2467 -0.1200  0.2124  0.4016
#> -1.4266 -0.6398  0.9005 -0.1684 -0.5092 -0.4786  0.1040 -1.3984 -0.0545  0.1077
#> -0.8864  0.2498  1.5034  1.7169 -0.9136 -0.5180  1.8250 -0.7705 -0.3491 -1.4549
#>  0.7045 -1.4276  1.2458  0.0681  0.2841 -1.1509  0.6430 -0.0092  2.4903 -0.9948
#> -1.9979 -0.2146  0.0795 -0.7446 -0.8069 -1.0800  0.3458 -1.2024 -0.3191 -1.0419
#>  1.4518  0.9924 -0.0946  1.3104 -2.3763  1.4113  0.9257  0.4154 -0.2077 -0.6819
#> -0.8661  0.7582  0.6981  0.6265  0.3230  0.5555 -0.8057 -1.1370  0.9123  0.4131
#>  0.2996  1.3826  0.0506  0.6029  1.7363  0.2499 -0.3363 -1.6712 -1.5648  0.0216
#>  1.1323  0.3088 -0.3450  0.3968  0.1468  0.1779  1.7171 -1.3381 -0.7658 -0.1068
#>  1.1395  0.7101  0.7285 -0.2548  0.0143 -1.8498  0.9480  0.6759  1.6773 -0.0160
#> -0.3222  0.2624 -0.4231 -0.5596  0.0866  1.1174 -0.2954  2.2101  1.6728 -2.2799
#>  0.8564  0.6022  0.9157 -0.1568 -0.8185 -0.8393  1.0830  0.1674  0.5156 -2.3665
#> -0.9369 -0.5991  1.8026  0.4808 -1.8478 -0.4942  0.2035  0.3935  0.6725  1.9315
#> -1.0645  0.0085 -0.7694 -0.0262  2.1454 -0.7105  0.6676 -0.2551 -0.6459 -1.0725
#> -1.7036  0.2807  1.1372 -0.8840  0.1720 -1.2349  0.3297  2.7102  0.7611  2.1566
#>  0.2049 -0.8210  0.5426 -0.9998  0.3564  0.5238  1.8351 -0.7703  0.0733  0.5672
#>  0.8354 -1.3800  0.6734 -1.6400 -0.9209 -0.4162 -0.9833  0.7082 -0.4158  0.4676
#>  1.3495 -0.8609  0.5458 -0.6864 -1.2344 -0.0315 -0.2143 -0.3765 -0.8656 -0.5473
#> -0.7810 -0.2588 -0.0348  0.5083 -1.7271  0.7244  0.3515  1.7358 -0.3704 -1.3180
#> -0.1309  0.3441  0.5554  0.9980 -0.9627  1.9227  1.2067 -0.6459 -0.2487 -0.7664
#> -0.0533 -0.1975 -0.5976  1.2417 -0.7440 -1.6580 -0.5437 -0.0130  0.5824 -0.5306
#>  0.9521  0.6204  1.3222  0.6195  0.7876 -0.1165 -1.0324  0.3160 -0.0325 -1.1437
#>  0.9729 -1.1462 -0.7702 -0.6026  0.4502 -1.2610  0.2189  0.5767  1.1265  0.3718
#> -0.4119  0.7935 -0.2874  0.7731  1.4249  1.2382  0.7635  0.4876  0.5341  1.8692
#> -2.6693 -0.7576  0.1224 -0.6360  0.0560 -0.9439  1.0144 -0.7951 -1.8352  0.5864
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]