Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.0184  0.8678 -0.8826 -1.1067 -1.0777 -0.6625 -0.0980  1.2419  0.1759  0.9245
#> -0.6875  0.2745 -0.9136 -1.0759 -0.8243 -1.0552 -0.0813  0.2023 -0.6280  1.7764
#> -0.5932 -1.6529  1.0170  0.7465  1.3906  1.0001 -1.0964 -0.6050 -0.5880  1.1073
#> -0.9444  0.2705 -0.6022 -1.0040  0.7215 -1.1560  0.8293 -1.0296 -0.2617 -0.9726
#> -0.7395 -0.7392  0.3283  0.2255  0.2895 -0.5386 -0.0855 -0.3552  0.2673  0.4483
#> -1.5191  1.3661  0.2988 -0.1989 -0.0700  1.0266 -0.0253 -0.5297 -1.2700 -1.5569
#>  0.0340 -0.1024 -1.2416 -0.9519 -0.1637 -0.2112 -0.8398  1.9574 -1.2428 -0.3528
#> -0.5858 -0.9438 -1.9260 -0.4293  0.0681 -0.1571  1.7577 -1.1510 -1.0168  0.1076
#> -1.3115  0.6392  0.2170 -0.1126 -0.8919 -1.1704 -2.2442  0.0189  0.8829 -0.9541
#> -0.3754 -0.0716 -2.4278  0.9404 -0.7274 -2.2369  1.7621 -0.1902 -0.7277 -0.8725
#> -1.2571  0.7767  0.1869  1.5462 -0.6702  0.9068 -1.6357  0.5041 -0.8857  0.5462
#> -1.6317 -0.2759 -0.5443 -0.5704  1.0064  0.6007  1.7798  0.6062 -0.0059  1.4158
#> -1.2462 -0.9890 -1.1055 -0.7819 -0.2973 -0.8271  0.7197 -0.4404 -1.2475 -0.4149
#> -0.5957  1.2314  0.6380 -0.4572  0.2924  0.7435  0.9748 -1.4466 -1.3622 -1.0553
#>  1.1450 -0.6178  0.1184  1.4220 -2.4349  0.0561 -0.6592 -0.0901  1.8157  0.5422
#> -0.5394 -2.1614 -0.3030 -0.4181 -2.2574  0.4172  0.3222  0.5863 -0.4305 -0.4891
#>  0.9824  0.2932 -1.3196 -1.0396 -0.1183 -0.4642 -1.3019  0.1310 -0.1874 -0.6717
#> -1.5166  0.6642  0.5894 -0.4420 -0.3959 -0.7939 -1.0010 -0.3478  1.0873 -0.5981
#> -1.3318  0.2121  0.6333 -0.9094 -0.5219  0.8262 -0.7258 -1.2746 -0.7336  0.9123
#>  1.3251  0.0168 -1.0632 -0.0055  0.9870 -1.0370  1.2832 -0.5022 -0.8774  0.4699
#> -2.3866  0.0383  1.5316  0.0289 -0.6169  2.0875 -0.2438 -0.6015  0.0277  0.8872
#>  1.6948  0.9087 -0.1144  2.3371  1.0750 -1.5491  0.8834  0.9707 -0.2503 -0.6422
#> -0.2542 -0.0421 -1.4476 -1.1752 -2.0889  0.6923  0.7877 -0.9837 -1.0867 -1.2955
#> -0.4994 -0.1524 -0.3668 -0.3670 -0.6224 -1.2817 -0.1774  0.8032 -2.6065 -1.8035
#>  0.0798 -0.5854  1.1686 -0.8882  0.0436  0.7126 -2.2706 -0.9940 -0.3980 -1.0701
#>  0.3462 -0.9339 -0.4566 -0.1900  0.1484 -0.3644 -0.5833 -0.3067  0.8290 -0.4591
#> -2.0575  0.4051 -0.0059  0.0319  1.8033  1.8825  0.9914 -0.7511  0.4601 -1.1774
#>  0.6208 -0.5038 -0.7802  0.1039  0.1360  0.5510 -1.6997  1.0779 -0.9263 -0.7072
#> -1.8498 -2.4691  0.6141 -0.1693 -1.9661  1.7064 -1.4289 -1.9337  0.1190 -0.6432
#> -0.8138 -1.2211 -0.0102 -1.0351 -0.9347  0.5295  0.4230  0.2806 -1.5121 -2.1024
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]