Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0655  0.1468  0.0308  0.0514 -0.5705  1.1907  0.6022  0.7415 -0.5689  1.0715
#> -1.1388 -0.3939  0.8710 -0.3611  0.0990  0.0109 -0.6887 -0.8038 -1.1486 -0.1718
#>  1.4645 -0.0003 -0.0049 -0.6969  0.5723 -0.0620 -0.5222 -0.3013 -0.0584  0.5479
#>  0.4904 -0.6658 -0.6572  0.0486 -0.6511 -0.6616 -0.0519  2.5821  0.3183  0.0785
#> -0.7579  0.7672 -0.7280 -1.7013  1.1865 -0.4496  0.0790 -0.8336  0.3385 -2.5021
#> -1.7636 -0.0312 -0.1930  0.0458 -1.4235  0.6685 -1.5851  0.5732  0.4857 -1.3657
#> -0.2728  0.8844  0.4826 -0.9990  0.2723  0.0040 -1.3987 -1.4077  0.9389 -0.4487
#> -0.9284 -0.0618  0.5808 -2.4175  0.0032 -0.4517 -1.4396  0.4712 -0.3162 -1.4059
#> -0.5544 -1.0261 -0.2793  1.0439 -0.5019 -0.7078  0.3284 -1.1019 -0.8223  1.0018
#> -0.0265  0.4374 -0.0329 -0.4254 -0.7318  1.2458 -0.0572  1.1856 -0.5040  0.1954
#> -0.6844  0.0480  1.0292  0.8157  0.2968 -1.0236 -0.2737  0.0103 -1.2074 -0.5070
#> -1.0940 -0.9611  1.0691 -0.5089 -0.7925 -0.9840  0.4340 -0.7659  0.7465 -0.1916
#> -1.5557  0.4815  0.9028 -0.6178 -0.4060 -1.5942 -2.0955  1.7387 -0.6317 -1.3050
#>  0.7251 -1.7062  1.4615 -0.5090 -0.6522 -0.6393  0.0008 -0.6241 -2.8077  0.3456
#> -1.2228 -1.4649  1.9429 -0.8095  0.4771 -0.0862  1.1321 -0.4838 -2.2234 -0.0142
#> -1.6954  0.6694  0.8150 -1.4391  1.1125  0.9828 -0.0301 -0.1093  0.8043  0.3559
#> -0.2580 -0.3940 -1.0817  1.5219 -0.4401 -0.4354 -0.0916  0.3229 -0.4458 -0.1954
#> -0.7584  0.2297  0.5670  1.0623  0.2172 -0.2364  0.6049 -0.0256 -1.1243 -0.8470
#> -1.0064  0.9737  0.4265 -0.3559  0.6607 -0.4472 -0.4694 -1.0057 -1.0891  0.1490
#>  1.8382  0.2027  0.0072 -2.9040  2.3617 -0.2183  1.5747 -0.0893 -1.0256 -1.6341
#>  0.3238 -2.0586  0.2632 -1.4847 -0.3913  0.3252 -0.4735 -0.3219  1.5620  0.4932
#> -0.1582  0.1491  0.1103  0.6910  0.8050 -0.5020  0.2298  0.3614 -0.1715 -0.6188
#>  0.4023 -0.9902 -0.6571  0.3273  0.0651  0.2063 -1.1984 -0.1703 -0.8274 -0.0688
#>  1.2098  1.2215  0.0460  0.0638 -1.4338  1.9473  0.5301 -1.1580 -0.1894 -0.5989
#>  1.7941 -0.0305  0.6468 -0.4763 -1.4256  0.7903  0.4177  0.7912  0.5069 -0.4256
#>  0.2309 -0.2276  0.1750 -0.2102  0.0776  1.1751  1.0607  1.2383 -0.5390  1.5542
#> -0.4210  0.2231  0.3819  0.2535  1.0476  0.7372  1.6830 -1.6247  0.4618  0.6321
#> -0.0619  0.8284 -0.0192  0.0635  1.1653 -1.7156  0.0180 -0.5106  2.4497 -1.0868
#> -0.2277  0.0812  0.6558  0.0560  0.5873  1.3601 -0.3041 -0.4072 -2.1987 -0.3147
#> -0.4427 -0.8668 -0.5145  0.2838  1.0460  1.1023 -1.0330 -0.3737 -0.9608 -0.8326
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]