Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.2724  0.6064  0.0054 -0.4857  0.4448 -0.2035 -1.1128  0.7056  1.3824  0.3292
#> -0.9251  0.8069  1.5398 -0.6744  0.4051  0.8157  0.5971  2.1103  1.6745 -0.4845
#> -0.6671 -1.0235 -1.3219  0.8720 -0.0375  0.2151 -0.5297  0.3348  0.3712 -0.5916
#> -0.9529  0.3168  0.9580  1.0002  0.4224 -0.6820  0.7133  1.0194  1.0431  1.1088
#> -1.1377 -2.7024 -0.8852  0.8887 -0.4191 -0.7039 -0.5559 -1.6642 -1.6665  1.4959
#> -2.0001 -0.4712 -2.7572 -0.3418  0.5207  0.3235 -1.7171 -0.3688  1.6232  0.1140
#> -0.7459  1.2220  0.5049  0.8936 -0.1060  0.7699  0.2290  0.0019 -0.6011 -0.1994
#>  0.2408  0.5073  0.0978  0.9128  0.7356  0.8918  0.3340  1.4848  0.7858  1.0229
#> -0.4659 -0.6200  0.2387  0.4889 -0.3287 -0.2548 -0.7360 -1.5391 -1.7990 -1.5732
#> -0.2622  0.9921 -0.1102  0.4466 -1.4126 -0.5207 -0.3929 -0.6413 -1.6719 -1.5777
#> -0.7944  0.5666 -1.9928  1.4454 -1.2285  0.8070 -0.0413  1.0708 -1.1708 -0.6008
#> -1.0473 -0.1440  1.5974 -0.4336 -0.2113  1.1174  1.2505  1.3083  0.7483 -0.1908
#>  1.5063 -0.1222 -0.5169  0.6663 -1.3472 -0.5001 -0.9234 -0.0169 -1.7841 -0.7435
#>  0.8602  1.6950 -0.6233 -1.7678  0.5181  1.1576 -1.8515  0.2445  0.2198  1.2430
#> -0.6984 -0.1855 -0.3756 -0.9888 -1.4528 -0.6338 -0.4475 -0.6563 -0.6567  2.8982
#>  0.5520 -1.2996  1.7324 -1.1214  0.0515 -1.2868 -0.8525 -0.7625 -1.2564  0.4824
#>  0.8900 -0.6409  0.0777 -0.6806 -0.6218  0.7224  1.1466  1.2180  0.5353 -0.8270
#>  0.3426  0.1041  0.5833  0.2215  0.0914  1.5796 -0.1841 -0.1050  0.6319 -0.2506
#>  0.0833  0.4794  0.7677 -0.4898  0.2189  1.2867  0.2317 -1.7386 -0.1697 -0.6082
#>  0.5852  0.0246  1.1102 -0.6536 -0.0022 -1.9855 -2.1045  0.1038  0.7367  0.2179
#>  0.6812  2.0387 -0.0156  0.0409  0.4794 -1.1921 -1.0313 -0.1589  0.7726 -0.2331
#> -1.4892 -1.4463  2.0237 -0.2195 -0.7155  1.0392  0.6676  1.1295 -0.1452 -0.8420
#> -1.2636  0.3454 -1.7893 -0.1266  0.0613  1.0771 -0.3656 -0.8582  0.6415  0.3300
#>  0.0403  1.0359 -0.0779  1.1833  3.3100 -1.6091  0.2618  0.6068  1.2801 -0.5032
#> -0.6874 -0.7338  0.0607 -0.7465  0.5032  0.9018 -1.6512  0.4058 -1.2716 -0.3907
#>  0.2988 -0.2905 -0.1737 -1.7527  2.4211 -0.7708 -0.0214  0.4171  1.0346  0.4130
#> -0.3092 -0.3130  0.1602 -0.7379  1.8245  2.8347 -0.1397 -1.3595  0.1689  0.1565
#>  0.1146 -0.8740  1.2358 -1.2879  1.3619 -0.1622 -0.1124 -0.7712 -0.1669 -0.9691
#> -0.9083 -1.8809 -0.3002  0.2487  0.3974  0.7617  0.1579  1.6501 -0.8748 -0.2844
#> -1.1631  1.4024 -0.6936 -1.4370  0.0084 -0.3676 -1.4345  1.0845 -1.6233 -0.2720
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]