Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.6848  2.0927  1.8617 -0.2107 -1.1132  0.8540  1.1947  1.8288  0.0424 -0.8323
#> -0.1491  1.6252  0.8867  0.3369  0.2630  0.2256 -0.0498  1.8990  0.3830 -1.2919
#> -0.6429  0.3543  0.3893  0.7148  0.3431  0.8842 -1.2070 -0.5581  0.5590  0.5058
#> -0.4644 -0.3579 -1.7990 -1.4921 -0.8514  1.3353 -1.1823  1.3318 -0.9123  0.2044
#> -1.2097 -0.3001 -0.0957  1.4339  1.3508  0.9625  0.7648 -0.9785  0.9769  1.1964
#>  0.8663 -1.4168  0.8531  1.4620 -1.3325 -1.0830 -0.1372  0.4696 -0.8019  1.9639
#> -0.9539  0.6221 -1.1673 -0.7614 -0.8468 -0.7873  0.5974 -0.1142  0.0340 -1.7044
#> -0.1992 -0.4974 -0.6880 -0.4588  0.4595 -1.2242  0.4182  0.9046 -0.8258 -0.1923
#>  0.3232 -0.5288 -2.1809 -0.2324 -0.5843  2.1100  0.3817  1.9227  1.2137 -1.7911
#>  0.5042  0.3840  1.1000 -0.2208  0.3872  0.5826  2.0990  0.4310 -0.4201 -1.0059
#>  0.9543  1.5803 -0.6023 -0.9078  1.5507 -1.0676  0.4747 -1.2731  0.5287  0.8672
#>  1.0733 -0.4126  1.2586 -0.6745 -2.4899  0.4348  0.0906  0.7521  0.6307  0.7659
#>  0.0882 -0.9236  3.4939  0.2117 -1.2133 -0.1257 -0.1868  1.2579  0.9444  1.7034
#> -0.7629  0.2493 -0.3114 -0.8081  0.5058 -0.1923 -1.5348  0.7390 -0.2150 -0.3600
#> -1.1690  0.4686  0.4997  1.0195  0.1269  1.0985  0.7716 -0.3321  2.1857 -0.6709
#>  1.0406 -0.8000 -0.2673 -0.4029  1.7309  0.6025  0.9819 -0.6156 -0.0502 -1.3232
#> -1.2829  1.4147  0.2076 -0.1388  0.0488  1.0897 -0.2491  0.4877 -1.0035  0.4201
#> -0.5145 -0.6971 -1.2520  0.1640 -0.3812  1.6128 -0.8703  0.5165 -0.7084 -0.1255
#>  0.2967  0.3417 -0.2400 -1.4330  2.2564 -1.0842 -1.1925  0.5499 -0.1943 -2.1065
#>  1.6605 -0.1924  0.4510 -0.8622 -1.3667 -0.6432 -1.5185 -0.1590 -0.5796  1.3018
#>  1.5462 -0.7354  0.2338 -0.5093 -0.8072  0.2711  0.2234  0.0705  1.2984  0.1934
#>  0.2737 -0.2660 -0.6050  0.2472 -1.8311 -0.8741 -1.9022 -0.2395  2.9239  0.9966
#> -0.2614 -1.5306  0.2646 -0.5869  1.1558 -0.3821  0.5723 -0.9354 -0.1428  0.5785
#> -0.8542 -0.0346  0.7419 -0.0619  0.1634 -0.3542  0.4993 -0.1010 -1.3145 -1.3641
#> -0.2555  0.3632  0.6872 -0.0830 -0.7597 -0.3735 -1.9139 -0.5871  0.9056  0.3490
#>  1.3307 -0.4173  1.5657  0.3177  0.6947 -2.0478  1.5374  0.3032 -0.6174 -0.7282
#>  0.3902 -0.7188  0.8856  0.1996 -1.2957 -1.7728 -0.3129  0.9756 -0.4734  0.5760
#> -0.1817 -0.3416  0.4857  0.0773 -0.4866 -1.3151  0.3022 -0.4075  0.9158  0.2220
#> -0.9231  1.3977 -0.3458 -0.5616  0.6268  0.2732 -0.7091 -2.5205  0.5058  0.8977
#>  1.1354 -1.7955 -0.1124  0.9349 -0.2104  1.2979 -0.3619 -0.6616 -0.1616  1.3762
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]