Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.5616 -0.4169 -0.4788  0.4149  0.7657  1.1631 -1.8826  0.3104 -1.0999  0.3128
#>  0.0917 -0.7233  0.2346  0.5104  0.1875  0.2969  1.1244 -1.4170  0.0693 -0.4345
#> -1.1412 -0.7564 -1.1766  1.2568  0.0944 -0.3067 -0.1186 -0.4424  0.3716 -1.0578
#>  1.9197  1.4874 -0.1832 -0.6273 -1.4155 -0.9541 -0.7215  0.4714 -1.7481  0.6431
#>  0.3645  0.7731  0.9949  0.0167  1.0658 -1.4731  0.3369 -1.3660 -0.0306  0.0680
#>  0.5052 -0.6097 -0.3235  1.6660 -0.4406 -0.6990 -2.8477  1.2522  0.0995  0.4442
#>  0.2724  1.3828 -0.3514 -0.5622  0.2425  1.2532  1.0218  0.1363 -0.4036  1.0367
#> -1.0155  0.0233  1.5818 -0.3421  0.4049  1.0766 -0.3381  1.9425  1.7181 -0.3458
#> -0.3963 -1.9166  0.6429  1.6415 -0.4705 -0.2153 -0.2429 -0.9185 -2.0491 -2.4369
#>  0.1430  1.0105  0.6283 -0.3204  0.2563 -2.2692  0.8471 -0.6733  0.0756  0.6351
#> -1.5032 -0.0056 -2.2715 -0.5152  1.1911  0.7680 -1.0609 -2.0738  0.8602 -2.3845
#> -1.8046  0.1470  0.5250 -0.7064 -0.0023  0.9113 -1.9583  1.0346 -0.1657 -0.0555
#> -0.4768 -0.0122 -2.2314  0.5925 -1.5804  2.0515 -0.6468  0.5098  1.2995  0.2889
#>  1.4653 -0.8450  0.6725  0.1709 -0.0853 -1.0187 -0.6701 -0.4651  1.5955  0.1558
#>  2.4615 -0.2959  0.4525  1.4173  2.0525 -1.5207  1.8472  1.2794  1.7800  0.1537
#>  0.8781 -0.2441  0.3206  0.3466 -0.0455  0.5425  1.9195 -0.7974 -1.9222 -0.1498
#>  0.2151 -0.3015 -1.0877 -0.4436 -0.0404 -0.1600  0.2028 -0.4019 -1.2857 -1.0879
#> -0.8855 -0.3932  0.2513 -1.2912 -0.9545  0.5944 -0.6123 -0.8840  0.2578  0.5892
#> -2.8133  1.1991 -1.8941 -0.7946  0.1517 -0.1180 -1.1292 -0.4062  0.5136  1.0551
#>  0.7887 -0.5568  0.5072 -0.0438  1.2109  1.0163 -0.4227 -1.0783  1.5999  0.4382
#> -1.0679 -0.4773 -0.0230 -0.1185  2.7177  0.6791  0.3665 -0.0270 -2.0001  1.2704
#> -2.3543  2.1000 -0.2435  0.6185 -0.5586 -0.3822 -1.3936  1.2534 -0.0392 -1.0085
#>  1.1859  0.5926  0.7800  0.3829  0.0664 -0.7454 -0.3169 -0.6771  1.0340  0.8221
#>  0.4615 -0.5605  0.1542  1.6260 -0.4423 -0.2672  0.0354  0.3990  1.6179  1.7501
#>  0.6962  0.1298 -0.0554  2.0029 -0.5725  0.0174 -1.1483  0.1318 -1.4132  0.8355
#>  0.3177 -1.7252  0.5382  0.8401  0.3904 -0.2261 -0.1575 -1.3791 -0.4175  1.3899
#> -2.6990  1.7281 -0.4941 -0.6267  1.5366  1.3889 -0.8845  0.3397 -0.9187 -0.2916
#> -2.0475  0.8586 -0.9220  0.3445  0.3066  0.0067  1.3715  0.3129  0.4720 -0.9943
#>  0.6445 -0.6282  0.9135  0.6033  1.0272  1.2313  0.0460 -1.3426 -1.3652  1.5586
#> -0.4197  0.4508  0.2315  1.5772 -0.6613 -0.6957 -0.5659  0.1810 -0.7063  0.2250
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]