Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 3.7777e-01 -5.4872e-01  2.2911e+00 -8.6273e-01 -2.6777e-01 -4.1112e-01
#>  7.9156e-01  3.6716e-02 -1.4898e+00  8.3361e-01  5.7618e-01 -1.6387e+00
#> -1.1227e-01  1.8358e-01  1.6820e+00 -1.5552e+00  1.1847e+00  5.6496e-01
#>  2.7852e-01 -1.2424e+00 -6.4066e-02  1.0775e+00 -6.1008e-01  9.7903e-01
#> -1.6206e+00 -3.8405e-01  6.2189e-01 -1.2932e-01 -1.2350e+00  3.0656e+00
#> -1.8594e-01  1.0977e+00  5.9281e-01 -1.0184e+00  6.7158e-01  1.5771e-02
#>  1.4640e+00  2.1956e-03 -2.4493e-01  3.9887e-01  3.8015e-01 -1.3487e+00
#> -7.7922e-01 -1.7693e-01 -1.9783e+00 -1.5332e+00  2.3852e+00 -5.1648e-01
#> -2.0152e-01  1.2269e+00 -1.2235e+00  2.0421e-01 -7.0449e-01 -2.2763e-01
#>  5.1346e-01  1.5169e+00  3.9040e-01 -1.2863e+00 -9.4989e-01 -1.3344e+00
#> -9.1929e-01 -1.2114e-01 -1.5403e+00 -1.3972e+00  2.1008e+00  1.1595e+00
#>  6.4680e-01 -1.6031e-01  2.3058e+00 -8.0874e-01  6.3515e-01  6.2412e-01
#>  1.6291e+00  1.0308e+00  6.4968e-01 -4.0686e-02  9.4252e-01  1.1555e+00
#>  8.5547e-01  1.8452e-01 -1.0507e+00 -1.6898e+00  1.5258e+00 -3.3401e-01
#>  5.6221e-01  1.0496e+00 -3.6535e-01  7.3633e-01  1.1323e+00  7.4629e-01
#>  1.3432e+00 -1.2365e+00 -1.3596e+00  9.9876e-01 -5.1792e-01  2.0764e-01
#> -3.7674e-01 -1.0283e+00 -1.7199e-01  2.3605e+00 -4.1129e-01 -1.7649e-01
#> -1.4846e+00  1.1431e-01  3.7655e-01  9.7820e-01  5.3043e-01  1.3671e+00
#> -3.3059e-01 -1.8585e+00 -4.8804e-01  3.9959e-02 -2.0862e-01 -8.8778e-01
#>  1.5087e+00  7.2937e-01 -9.5982e-01  1.7872e+00  6.4811e-01  1.0598e+00
#>  1.1967e+00 -1.8081e+00  1.1610e-01  8.9897e-01 -4.6904e-01  4.9142e-02
#> -1.0567e-01 -7.6987e-01 -2.3754e-01  5.2149e-01  1.8307e+00  6.2576e-01
#> -7.4015e-01 -7.0611e-01 -1.1991e-01  7.9818e-01 -1.4061e+00 -3.1607e+00
#>  8.5013e-01 -1.0625e+00 -3.2137e-01  7.7433e-01 -4.9830e-01  9.7868e-01
#> -2.0170e-02 -1.0860e+00 -4.1689e-01 -3.2634e-01 -2.4543e-01  1.2784e+00
#>  1.0572e+00 -5.9613e-01 -9.1787e-01 -6.9673e-01 -1.6970e+00  4.1112e-01
#>  1.9107e+00 -6.7303e-01 -1.3950e+00 -1.8950e+00  4.2887e-01  1.4012e-01
#> -1.6913e+00 -1.8179e+00 -2.1202e-01  7.1827e-03  2.6872e-01  2.8531e-01
#>  1.4430e+00 -5.6322e-01 -1.6728e+00 -4.5787e-01 -6.5419e-01  8.8595e-01
#>  5.2011e-01  1.5335e+00 -5.5979e-01  3.4163e-01 -8.1677e-01 -8.8997e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]