Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 1.5096e+00  1.3116e-01  1.1392e+00 -4.5636e-01 -7.5481e-01 -1.2280e-01
#> -1.8363e+00 -2.1429e+00 -7.2248e-01 -2.2856e+00  3.0650e-01 -1.3110e+00
#>  1.9874e+00 -3.6878e-01 -2.6892e-01  1.1386e+00 -7.1595e-01 -1.2085e+00
#>  1.5460e-01  2.3883e-01 -6.2957e-01  8.7585e-02  1.3150e-01  6.1881e-01
#>  3.3207e-01 -4.5859e-02 -9.9796e-01  1.1775e+00  1.0336e+00  1.5800e+00
#>  7.0525e-01  6.0894e-01  1.2154e+00  5.4468e-02 -1.1802e+00 -8.1364e-01
#>  1.2446e-01 -7.8349e-01 -9.3327e-01  2.1545e-01  2.3245e-01  1.1281e+00
#> -1.0239e-01 -1.0246e-02  7.2782e-01  2.8946e-01 -4.5057e-02  1.3112e+00
#>  1.1639e+00  7.7890e-02 -2.7355e-01 -3.4030e-01  2.8782e-01 -3.8626e-02
#> -5.7892e-01 -6.0822e-01 -4.4836e-01  1.9124e-01 -1.1212e+00 -1.0188e+00
#> -1.0138e+00  1.1420e+00  6.1265e-01  1.8015e+00 -9.3384e-02 -8.5963e-02
#>  1.2014e+00 -7.1360e-01  1.0857e+00 -8.6639e-01  7.9055e-02 -1.9369e+00
#>  7.1708e-01 -5.5932e-01  7.0231e-01  1.3263e+00 -1.1959e+00  4.1676e-01
#>  8.1677e-01  9.3566e-01  1.6068e+00 -2.9382e+00 -1.0916e+00  4.2248e-01
#> -4.8566e-01 -2.8661e-01  5.3124e-01  3.8754e-01  3.0417e-02  1.5466e+00
#>  2.6389e-01  1.5934e+00 -1.6420e+00 -6.5159e-01  9.5660e-02 -6.7226e-01
#>  6.8503e-01  5.0275e-01 -1.2630e+00  1.6050e-01 -1.8231e+00 -7.7959e-01
#>  9.9642e-01  4.0588e-01 -3.6457e-01  3.9711e-01  2.3429e-01 -8.9865e-01
#>  8.5652e-01 -3.9823e-01 -1.7846e+00  2.2930e-01 -7.0907e-01 -1.4270e+00
#>  6.7130e-01 -1.8245e-01 -1.7087e+00  2.2631e-01  1.2954e+00 -5.1140e-01
#> -2.2064e+00 -1.0970e+00  1.5351e+00 -9.7493e-01 -9.5224e-02 -1.1817e+00
#> -7.1948e-01  1.7413e+00  6.8303e-02 -5.3909e-01  1.9382e+00  1.9530e-01
#>  1.3080e+00  1.9825e-02  3.9052e-01  3.3541e-01  1.0223e+00 -9.7263e-01
#>  4.0751e-01  2.5105e+00 -2.7084e-01  1.0240e+00 -1.5552e+00  1.6737e+00
#>  4.7725e-01  5.6593e-01 -8.1805e-01  1.9978e-01  2.8802e+00  2.8731e+00
#>  1.3630e+00  1.2483e+00  5.7665e-01 -8.4149e-01 -1.3883e+00  1.8739e-01
#>  3.4029e-01  1.7885e+00  3.8060e-01 -5.5664e-01 -1.3949e-01 -9.2923e-01
#> -9.8056e-06  3.6409e-01 -6.1332e-01  4.4544e-01 -3.8125e-01  3.3481e-01
#>  1.6706e+00  4.7056e-01  1.3781e-01 -7.8729e-01  1.8820e-01 -1.1446e+00
#>  6.4843e-01  5.5328e-01 -9.9356e-01  1.6082e+00 -1.3883e+00  1.1317e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]