Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 2.1151  0.4732 -1.6700  0.1571 -1.2761 -0.2848  0.9293  0.9274  0.6807 -0.2939
#> -0.9872  1.5253 -1.3633  0.0457  0.4805 -0.0778  0.0454 -0.7475  0.5942  0.9424
#>  1.3295  1.1509  2.4193 -0.6442  0.2142  0.5324  0.0073  2.0181  0.5334 -0.6573
#>  1.0455 -1.4330 -1.9833  0.1712 -0.7668  0.1847 -2.7583  0.8268  0.9431 -0.6875
#> -0.4623  0.6328 -0.3488 -1.0819  1.5084  1.1920  1.1509  1.4849  0.4203  1.1867
#> -2.0253 -0.2208  0.9676 -0.1386 -1.1844 -0.4378  0.4814 -0.1039  0.2177  1.2009
#> -0.6496 -0.6281  0.8377 -0.3573 -0.4102 -0.4452 -0.2144 -0.7876  0.8872  0.1300
#> -0.0531  1.2671 -2.1277 -1.1653  0.4269 -0.8912  0.3155  0.5703  0.9428  0.7616
#> -0.3403  0.9900 -0.1562  0.1310 -0.3279 -0.7072  0.6322 -0.1000 -0.3962 -1.0567
#>  0.0597  0.8856 -0.0225 -0.2025 -0.8548  0.0276 -0.1821  0.6677 -1.0576  1.5825
#> -0.1524  0.2491  0.6373 -1.6284  0.3666 -0.0546  0.9371  0.5882  1.4424 -1.2401
#>  1.9938 -2.5959 -0.6911  0.9042 -1.8720 -1.6406 -0.5178  0.5700 -0.0837  0.0281
#> -0.8241  0.4853 -0.9656  0.9484 -0.8251  1.4815  0.4044  0.2443  0.2204 -0.1150
#> -0.5713  0.6904  0.1652 -0.2407 -1.0203 -1.0252  0.4704  0.3392  0.8146 -2.0341
#>  0.9393 -0.2379 -0.1765  0.8155  1.4709 -0.2617 -0.9432 -1.1837 -0.2147  0.8569
#> -0.4967 -0.1533  0.4826 -0.8976  0.9742  1.4741  2.0674  0.7844  0.5915  1.0535
#>  1.6069  0.6656 -0.2974  0.9864  0.3262  0.4165 -0.5272  0.4030 -1.1443  0.2858
#>  1.0939 -0.6829  0.3945 -0.0989  0.3688 -0.4372  0.8690 -0.3378  0.5832 -0.4821
#>  1.4697  1.9260  1.3245 -0.6588 -1.4331  1.0683 -1.5484  0.3571  1.6917 -1.1177
#> -0.5554 -1.2869 -0.0480 -0.6227  0.3325  0.4489 -1.1902  1.0108  0.5961  0.6904
#>  1.3671 -0.4091  0.0912 -0.3169  0.6272 -0.4775 -0.6348 -0.4947  0.5638  1.1902
#>  0.4783 -0.0299  0.1599 -0.5570  0.5792 -0.2100 -0.2111  0.9431  1.4757 -2.2189
#>  0.0021  0.4763  0.5771  0.9141 -1.0484  1.0153 -0.3380 -0.9902  1.2545 -0.2573
#>  0.2869  0.5372  0.4020 -0.2981 -0.0541  0.9671 -0.3510 -0.6788  1.3667 -1.0485
#>  0.2508  2.0081  3.0456 -0.0423  2.4647 -0.7405 -0.8544  0.4841 -0.3590 -0.2291
#> -0.6972  0.9688  0.0737 -0.5855 -0.2815 -1.5344 -1.1530  0.1976  1.4164  0.4010
#> -0.2489  1.3584 -0.7848  1.5920 -0.6127  1.7371 -0.9679 -1.0659 -1.9616 -1.0200
#>  1.0870 -0.8484 -0.0757 -0.4397  0.6173  1.9376 -1.1421  0.1376 -0.5468  0.0630
#>  2.0194 -0.4344  0.8881  0.5970  0.3560  1.4961 -0.6336 -0.6478 -2.1245  1.4913
#> -0.1324 -0.4888  0.8321  0.1021  0.4506  2.6036 -0.0392 -1.4465 -0.5384  0.4048
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]