Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5687  0.0070  1.4612  0.6909 -0.4522 -0.3684 -0.4142 -0.5019  0.9227  1.8714
#> -1.0869 -0.1458  0.0257  0.1696  1.4234 -1.1335 -1.1145  1.5461  1.0660 -0.7718
#>  0.3069  0.9892 -0.4290  0.0838  0.5975  0.7229 -0.2051  0.3498 -0.1718 -1.1970
#>  0.1615  0.2533 -0.0179 -0.5301 -0.7736 -1.0384 -1.3627 -0.3990 -0.5161  0.0096
#> -0.0125 -1.3459  0.4530 -1.4870  0.1550  0.9065  1.6210  1.0615  1.1201 -1.2960
#>  0.8493  0.6954 -0.3410  0.4530 -2.9757  0.3997 -0.2643 -0.8797 -0.8028 -1.1631
#> -0.0624 -0.1642 -0.9117 -0.9289  1.8578 -1.8962 -1.4428  1.6503 -2.0188  0.4714
#>  0.9035 -0.1087 -0.7281 -1.0794 -1.5898 -0.0129 -0.7935 -2.0200  0.2292 -0.3527
#> -0.3899 -0.0424 -1.1545 -0.4833 -0.0317  0.7122  0.5521 -0.3338 -0.4320  1.2315
#>  0.2758 -0.6347 -0.1488  0.6300  0.5194 -1.4664 -1.5500  1.7694  0.6068 -0.1526
#> -0.7650  0.7439 -0.5913  1.1632  0.0217  0.6691  0.7544  0.4426  0.5216  0.9565
#>  2.0122 -0.6329 -0.0728 -0.1957 -0.5163 -0.8047 -0.0658  2.0984 -1.0279 -1.1607
#>  0.5966  0.5162  0.3633 -0.3015  1.1581 -0.5069  1.8065 -0.0354 -0.2438 -0.7287
#>  1.1025  0.4179 -1.1014 -1.3617 -0.3353 -1.2740 -0.3178 -0.9154 -0.0322 -0.5555
#>  0.3468 -1.7997 -1.5714 -0.0128 -1.1163 -0.1238  0.2243  0.0462  0.4679  0.8213
#>  0.6751 -0.0151 -0.5547 -0.1215  1.6907  0.6899  1.0653 -1.0551  0.1429  2.0132
#>  0.4145  0.0439  0.5686 -0.6732 -0.0769  0.1918 -1.6808  0.2238 -2.0255  1.0007
#> -0.9152 -2.1958  0.2509  0.9211 -0.2939 -1.5530 -0.8359  0.5878 -0.9856  0.4136
#>  0.1882  0.7581  0.2660  1.4298 -0.3433  0.3911 -0.5040  0.9074 -1.9941 -0.1682
#>  0.0289 -0.2009 -2.4617  0.3572  1.1746  0.1379  0.4037 -0.3596 -0.3904  0.2948
#> -0.1328 -0.1234  0.5261  0.5787  0.3540 -0.0296 -1.0245 -1.0088 -0.0765 -2.5832
#> -0.8634  0.2713 -0.2872 -3.2390  0.8761  1.4438  0.9250  0.7624 -0.7953  0.0915
#> -2.1674 -1.1404  1.1881 -1.0898 -0.4347 -0.7526  1.1190  1.6640 -0.4028 -0.0652
#>  0.5883 -0.7279 -0.0353  0.7457 -0.2960 -0.5379 -0.8667  0.4193  0.6712  0.3281
#> -0.1390  1.7902 -0.0367 -0.3358 -1.5040 -0.4363 -0.4279 -0.5382  1.7074 -1.0922
#>  0.0108  0.1390 -0.2363 -1.7228 -0.5376 -0.0229  0.7272 -1.8379  0.8019 -0.9710
#> -0.5677 -1.7778  0.2487 -0.7216  0.0907  1.4559  0.2274  2.0772 -0.2528  0.0789
#>  1.4749  0.5494  0.3045  0.4300 -0.6468  1.2304  1.0722 -0.7660  1.5355  0.6535
#>  0.4316 -0.4002  0.8414 -1.0281  1.3488  1.2060  0.6426  0.5919  0.9968  1.0941
#>  1.6390 -0.1746 -1.0546  0.4412 -0.8430  0.4203  0.4415 -0.3792 -0.6580 -0.0358
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]