Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5606  1.0046 -1.4301  0.4761  0.6515  0.3934 -0.0529 -0.8896 -0.4941 -0.2134
#>  0.5490  1.8688  2.4371  0.4868 -0.9704  0.4602  0.1366  1.0186 -0.6390  2.2259
#> -0.1760 -1.7394 -0.1953 -0.3875 -0.5175  1.2250  0.9097  0.0033 -2.2271 -0.5083
#> -0.0794  1.9879  1.7336 -0.9711  0.1124 -0.3011  0.1063  2.3840 -0.9189 -0.8202
#>  0.9202  1.0293  1.3277 -0.2904 -0.1230 -0.3062 -0.2370  0.5168 -0.6570  0.9576
#> -0.5774 -0.4628  0.6944 -1.1373 -0.2539 -1.2455 -0.7064 -1.1285  1.1499  1.4572
#>  0.4530 -0.2447  0.1911 -0.0745  0.9713 -0.4247  0.3684  1.7852 -0.3538  0.0150
#>  1.0892 -0.3831  0.5816 -1.8438  0.2929  1.9695  0.8641 -0.0555  2.0036  1.2313
#>  0.2546  0.4427  0.4746 -1.6547  0.8735  0.0085 -0.2324  0.1241  1.6457 -1.4212
#>  0.2915  0.2801 -1.3944  0.1856 -2.3297 -0.5467  1.0602 -0.4761 -1.3665 -1.2885
#> -0.8161  0.4642  1.2733  0.0326 -1.2469 -1.2896 -0.4668 -0.7277  1.1568 -0.5447
#> -0.4139 -1.3532  1.1526  0.7975 -0.5884 -1.3702  0.0755  1.0150 -0.1335  0.0908
#> -0.1624 -0.5046 -0.0714 -0.8521 -1.8152  0.7219 -0.3049 -0.4485  1.1050 -1.9959
#> -1.0927 -0.6417  1.3295 -1.3496 -0.8063 -1.0761  0.6782  2.1168 -0.8509  0.4212
#> -1.1549 -0.5068 -0.1810 -0.6138 -0.3794 -1.6759 -0.6709  1.1290  0.5771 -1.0184
#> -0.3816 -0.1554  2.4357 -0.6125  0.9646  0.2788 -0.6752  0.9641 -1.2883  0.6286
#>  2.5444 -0.7443 -0.8808  1.2164  0.3833 -0.5932 -1.3053  0.0415 -0.1236  0.2516
#>  0.7012 -2.1062  0.4204  2.4385  0.3876  0.3966 -0.6053 -0.7740  0.2119  0.9025
#>  0.2644  0.5971  0.5162 -0.4118  0.6788  0.9088  0.6218 -0.0924 -0.0054  0.3318
#>  0.8879 -2.0708  0.2877  0.0285 -0.7850  1.0092  0.4886 -0.9990  0.1865 -0.2024
#>  1.3942  0.0122  0.8669 -0.2709 -0.7917 -0.9725 -0.2627  0.5187  0.4453 -1.3777
#> -0.4822  0.9722  0.1676  0.5381  0.0718 -0.2459  1.2511 -1.2056  1.0258  0.2386
#>  0.2024 -2.7770 -1.3710 -0.3372  1.3907  0.3598  0.9825  0.0038  0.7075  1.0358
#> -0.6175 -0.0759 -0.0376 -0.6393  1.0665 -0.1982  0.2182  0.1002 -0.0291  0.9479
#> -0.3965  0.5761 -0.7403  1.4519 -0.3455 -0.2597 -0.5990  1.1917  0.9053 -0.9339
#> -0.8126  1.8531 -0.3307 -0.0531 -0.0471  1.7449  0.2022  0.7252  2.3689  1.7347
#>  1.3805 -0.6511  0.5066 -1.1401  0.0733 -0.3660  0.4690 -0.0377 -0.2858 -0.9839
#> -0.2645  1.2210  1.3924  1.2425  0.6915  1.0607 -1.1142 -1.2871  1.1822  0.6081
#> -0.0509  0.7424  0.2485 -1.0413  0.9571 -0.8585 -0.7799 -1.1300 -0.4999 -0.2893
#> -0.1758  0.4445  0.9324  0.6432 -0.4093 -0.8280 -1.0471  0.1065  0.5857 -0.3341
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]