Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5571 -0.8651 -0.5948  2.1702 -1.5050  0.7836  0.2937 -0.0738 -0.6079  0.5763
#>  0.6295 -0.5052 -1.3774 -2.2958  0.8031 -1.6355 -1.2236  2.1059  1.2720  0.5700
#>  0.2475  0.6753  1.0708 -0.8991 -1.4031  0.2143  0.4901 -1.1078 -1.1304 -0.0363
#>  0.8792 -0.8068 -1.4099  0.7937  3.0773  0.3176 -0.6961  0.4371  0.7423 -2.0600
#>  0.3072  2.9003 -1.8827  1.2867  0.7334  2.1646 -2.5665  0.4018 -0.4668 -0.0842
#> -0.4548 -0.6868  1.4644 -1.2383  1.2635 -0.3759  0.0740 -0.4356  1.2755 -0.5823
#>  0.4475 -1.3677  0.3605 -0.9641  2.0984  0.1811  1.0954 -0.0241  1.5264  0.3929
#> -0.3057 -0.0275 -0.2529  1.2839  0.7940  0.4005 -0.5389  0.2001  0.7501  0.3769
#>  0.8812  0.0047  0.3560  0.0233  0.1130 -0.1915  2.1073  0.5257 -0.4681 -1.6129
#> -1.0216 -0.8964 -1.6086  1.3115 -0.2803 -0.8714  0.5529 -1.3722  0.6614 -1.6382
#>  1.0749 -0.5522  0.0137 -0.9210 -0.2503 -0.4913  1.5350  1.3006 -1.2035 -0.3267
#> -0.1590  0.7824  0.4276  0.3148  0.3647  0.8394 -0.6273 -0.2389  0.1779  0.2634
#> -0.1001  0.4129 -0.5478 -0.3017  2.1491  1.1088  1.0739 -1.4482  0.1632  0.3402
#> -1.7881  0.7846 -0.7602  0.6789  0.4354 -0.1046  0.6851 -0.6035 -0.6943  1.2486
#>  1.5235 -1.1849 -0.3640 -1.5583  0.3242  1.0069  1.0980 -0.2119  0.4187 -0.8735
#>  0.7953 -1.4562  0.3942  0.5070 -1.1212  0.6772 -0.1866  0.8010  0.1095 -0.3010
#> -0.1239  0.3079  1.2829 -1.9586  1.4544 -0.4115  1.0659 -0.3058  0.0838  0.3182
#> -0.5489  0.0207 -0.7045 -0.4707  0.2156  0.3147 -0.7581 -2.3197  0.0917  0.2543
#> -0.6470 -0.2153 -1.8045 -0.5292  2.0347  0.8746  1.6484  1.2223  1.0118 -0.0435
#> -0.3802 -0.0234 -1.0095 -0.6327 -1.6333  0.0184  0.7332 -1.2988 -0.8406  1.5720
#>  0.0740  2.0069 -2.0481 -0.0699  0.8859 -0.0681 -0.0997 -1.5680  0.8241 -1.0356
#>  0.1901 -0.1488  0.2863  0.5150  1.2847 -0.4899 -1.4707 -0.7271 -1.2426  1.0677
#>  0.1838 -1.6664 -1.6857  0.8232  2.0589  0.4072  0.5853 -1.4916  0.8454 -1.0580
#>  0.2054 -0.4124 -0.2254  0.9446  1.6427  1.6465  0.3151  0.1261 -0.2543  0.3974
#> -1.1071 -0.0711 -1.3916 -1.7614 -0.3304  0.9522  0.0254  1.1270  1.1425  1.3566
#>  0.0042  0.2920 -1.5104  0.6815  0.7473 -1.2801  0.0912 -0.7751 -0.2115 -0.5780
#> -0.4120  0.8770  0.0189  0.9835  0.6499  1.5658 -0.4608  1.0146  0.8933 -1.5730
#> -0.9964 -0.5917  0.9217 -0.3805  0.1916 -0.3026 -0.2041 -0.8768 -0.2928  0.3986
#> -0.7377 -2.1551  0.3908  0.0610  0.0025 -2.6566  0.7335 -0.6115  0.9247 -1.5756
#> -0.5260 -0.5964 -1.6456  1.2862 -0.1976 -1.3059  0.5663 -0.2775  0.3955 -1.1529
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]