Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.9421  1.1687  0.3858  2.0517  0.8848  0.1362 -0.2804 -0.1083 -0.1843  0.2245
#>  1.9008  0.1935  0.7493  0.4777  1.1058  0.5888  0.0581  0.4011  0.6114  0.9786
#>  0.9753 -0.0953 -1.0823 -1.4350 -1.2598  0.7980 -0.7587 -0.4069  0.4965  0.2105
#> -0.1528  0.6020  0.5692 -1.5875  0.7729  0.3572  0.3523  1.4877 -0.7014 -0.0373
#> -1.7881 -0.2421 -1.3950  0.5004  0.3996 -0.5259  1.3404 -2.3043 -1.1620 -1.6293
#> -0.3543 -1.8199 -0.0516  0.2623  0.0539 -1.6098 -0.5595 -0.7431 -0.3304  0.4993
#>  1.2783 -0.0617 -0.4284  1.5022 -0.0151  0.6231 -1.4225  0.5270 -1.2289  0.4551
#> -0.0479 -0.3662  0.8641  0.6780  0.2397 -0.0090 -0.0567 -0.4533 -2.0223 -0.8656
#>  1.3038 -1.1671 -0.8492 -1.5918 -0.3443  0.1572 -1.7095 -1.2375  1.5700  1.2279
#> -0.7861  0.4017 -1.6970 -1.2382  0.4236  1.2174  1.5760 -0.4437  1.2227  0.0335
#>  0.1775  0.2915 -0.4135 -0.5204 -1.5263  0.1835  0.5927  0.5972  1.2385 -1.5882
#> -0.6868 -0.8276  0.0094 -1.3069 -0.5914  0.4668  1.2526 -0.6688 -1.5904  1.3739
#>  0.7274  0.2273  0.3188 -0.2705  0.2342  0.2428 -0.1411 -0.6123 -1.0975  1.3795
#>  0.3005 -0.6311 -1.0246  1.5778  1.2108 -1.7726  0.9893  2.5152 -0.1590 -0.5827
#>  0.9926  0.8717  0.0876 -0.2030  1.0367 -0.0575  0.8841  0.5835  0.2253 -0.9040
#> -2.5824  0.3520 -2.8967  0.5340 -0.1098 -0.2585  0.4107 -1.0589  1.2069 -1.3364
#> -0.9661  1.2624 -0.2052  0.0890  0.6477  0.9046 -1.2798 -0.7228 -1.0065  0.8595
#> -0.0463 -0.0412 -0.5662 -1.0058  2.0381 -1.7426 -0.9928  1.3413  0.9273  2.3828
#>  1.2252 -0.0940 -0.9078 -0.1584 -0.2048  1.5401 -0.9729  0.4695 -0.4928  2.3519
#> -0.0851  0.6540  0.2808 -0.0043  0.5296  0.4558 -0.0944  0.4643 -1.6709 -1.5964
#>  0.1771 -1.9951 -0.7179  0.5502 -1.6495  2.2735  1.3810 -0.2673 -0.7911  0.7609
#> -0.5043 -0.4461  0.6446 -0.1060 -2.9611 -1.6359 -1.6866  1.1806 -0.3760  1.9109
#> -1.9579  0.7630 -0.8883 -0.5746  0.5535 -0.8929 -0.2249 -0.9375 -1.2054  0.5580
#>  0.1200  1.5024  0.5584  0.4202 -1.2101 -0.8611  0.9234 -0.6876  1.7266  0.8655
#>  1.5614  0.6998 -0.4152 -0.4910  0.3349 -0.2204  1.6566 -0.8266 -0.3173  0.8716
#> -1.2801  1.0227  1.0810  0.0924  0.0463 -0.3192  2.0945 -1.9725  0.6712 -0.2187
#> -0.1202  1.9907 -0.5600 -0.1873  0.6358  0.5844  0.9689  0.2568 -0.9057 -0.0308
#> -2.1143 -0.2304 -0.9861  0.3390 -0.2308 -0.7581 -1.1083 -1.5010  0.1681  1.1255
#>  0.7240  0.0951 -1.4718  0.1806  1.8900 -0.0523  0.5152 -0.1157  0.5735  1.6336
#>  1.6546  1.5936 -0.6547 -0.4408 -0.2122  1.5303 -0.0959 -0.1417 -0.6729 -0.2470
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]