Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-0.0009  0.0037  0.5204 -0.3667 -2.0531  0.2671 -0.5804 -0.2076 -0.6655  1.1677
#> -1.6947 -2.3617 -0.0151 -0.9661 -0.5683 -2.4593  1.0112 -0.6164 -1.1062  0.1686
#>  3.0387  0.7551 -2.3594  1.6931 -0.4003 -0.9425 -0.1783  1.5772 -0.4279  0.7089
#>  0.7899  1.6087 -0.0747 -1.7557  0.0622  0.3763  0.3831  0.5895  0.3669  0.0486
#> -1.3491 -0.1465 -0.0981 -0.1463  0.7424  0.5160 -0.3618 -0.5935 -0.2595 -0.6434
#> -0.7859 -1.3648 -0.0414  0.2196 -0.8855  1.3220  0.0348  0.9984  1.1103 -0.2425
#>  0.8881  1.0576 -0.4742 -0.4879 -1.1058 -0.6111  0.4100 -0.4836 -2.2143 -0.5398
#>  0.5408 -2.1837 -0.2428  1.4708  0.2433  1.0932  0.7066 -0.0637  0.9693 -0.9876
#> -1.5370 -0.3757 -0.9456 -0.5917 -1.4087  0.9152 -0.2538 -0.1560 -1.5692  0.3750
#> -0.1467 -0.0512 -0.3315 -0.2021  1.3404  0.7150  0.6986  1.8484  0.1063 -1.9778
#> -0.4032 -0.7702 -1.5078  0.2106 -1.3181  1.4029 -0.2071  0.6988  0.5814 -1.2226
#> -0.2139  0.2826 -0.6051  1.4271 -0.5023  0.1380  0.7487 -0.6512 -1.0667  0.6083
#>  1.1542 -1.1072 -0.0416 -0.7018 -0.2876 -0.5697  1.3263  1.3296 -0.7422  1.0843
#>  0.3823  0.3361  0.9064 -0.6069  1.0543 -1.0274 -1.0400 -1.9305 -0.0312  0.1556
#>  1.1169 -0.3283  0.1506  0.5543  0.8330 -0.6179  0.0503 -2.2701  0.7123  0.7061
#>  0.1040  1.7993 -0.8401 -0.0205 -0.0334 -0.9733 -1.3252 -0.4537 -0.3771 -0.5454
#> -0.6753 -0.3866  0.1658 -0.5853  0.8411 -0.6265  1.0690  0.8731 -0.9840 -0.0668
#> -1.3578  1.6471 -1.1438  0.1968 -0.1045 -0.2291  0.1634 -1.6033 -0.4936  0.3136
#>  0.3749  0.8226 -1.4063 -2.0339  0.8500 -0.2165  0.4049  0.3529  0.2333 -1.1750
#>  0.1571 -0.0908  0.3734  0.4290 -0.8236 -0.4525  0.6144  0.3749  0.0308 -0.1780
#>  0.1249 -0.0379  0.8394  0.0880  1.3442 -0.3450 -0.8468 -0.4069  1.3823 -1.0155
#> -1.7637  1.3288 -1.7400 -1.2977  2.1511 -0.6051  0.4648  0.7022  1.5900  1.9261
#> -1.2073  1.6272  0.8857  1.5837  0.8221 -0.0738 -0.8853  1.6199  0.0518  0.1564
#> -0.6355  0.8778  1.1055  0.6252  0.8287 -1.6925 -0.7142 -0.0719 -0.5059 -0.9019
#>  0.3946  0.4433  0.7126  0.2315  0.0714 -1.3102  0.5767  0.2772 -0.2313 -1.4692
#>  0.3674 -0.8891  0.0545 -1.2286 -0.3329 -0.8300 -0.0932  1.3368  0.1294  1.9112
#>  1.8811  1.1714  0.9356  0.0622 -0.1098  1.6281  2.4690  0.8353 -0.6021  1.1895
#> -0.9665  0.7614 -0.7413 -0.0969  1.6492 -0.8877  1.1068  1.9219  1.6268 -0.3664
#> -1.3160  1.0413  0.0369 -1.9015 -0.9639 -1.8736 -0.5516  0.0897 -1.4888  1.9580
#>  0.9566  0.2562  0.9344  0.5393  0.0169  0.8452 -0.8670 -0.6659 -1.0230 -0.2476
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]