Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.9121  1.7354  0.6261 -0.1344  0.0705  1.2428 -0.2857 -1.2040 -0.0267 -1.3586
#>  0.4684  0.1377 -0.2111  1.0757  0.3916 -1.2711 -1.0196 -0.3493 -0.6903  0.4606
#>  0.2911  0.2254 -0.7516 -0.0272  1.2340 -0.1374 -1.4144  0.4074  1.6609  0.4902
#> -1.1668  0.5908  1.4299 -1.4970 -0.7824  1.5409 -1.8090  2.7557 -0.8530  1.2697
#>  0.1869  1.1757  0.5170  2.0310  1.8831 -0.0949  1.2803 -2.2646 -0.5972 -1.7845
#> -1.7092  0.0662 -1.1192 -2.7235  0.0721 -0.1061  0.6671 -0.6404 -2.1541  0.9743
#>  0.0811 -0.4086 -0.5505  1.1461 -0.2274 -0.2854  0.7849  1.4388 -0.7569 -0.5886
#> -0.3760 -0.2488  1.6008 -1.3115  0.4288 -0.1826 -0.2699  0.1694 -0.1459  0.8248
#> -0.0577 -0.3182 -0.0894  0.3938 -0.0350  0.5239  1.5452 -0.7810 -0.3104  1.0347
#> -0.7033  0.1672 -2.2606  1.2900 -1.5984 -0.9105 -1.9556  0.2050 -0.4799 -0.3137
#>  0.2515  0.6319  0.3225 -0.6941  0.7364 -0.3899  0.2365  0.9034 -1.0604 -1.3505
#> -1.0687 -0.4525 -0.6593 -0.2376  0.2101  0.4025  0.1135 -1.8499 -0.3388  1.8180
#>  0.3358 -0.1575  0.9796  0.0785  0.6125 -0.5508 -1.0518  0.4646 -0.3252  0.5936
#>  0.1670 -1.6140 -0.4164 -0.7498 -0.5260  0.0235 -0.6084  0.2163 -0.3581  0.2614
#> -0.9433 -1.3393 -0.2485  0.7071 -1.2392 -0.6317 -0.8411 -1.3751  0.3082 -0.1359
#> -1.0242  0.0878 -2.2584  0.2480  0.3794 -1.1494  0.1006 -0.8941 -0.4752  0.2403
#>  1.3817 -1.4768 -0.4239 -1.3921  1.0450  0.1957 -0.2521 -0.7233 -0.0468 -0.3235
#>  1.2344  1.0038 -0.6826  0.9396  1.3536 -0.8654  0.7480  0.0689  1.2024  0.2341
#>  0.7143 -0.2342  0.7961 -1.0645  0.0841 -0.5419  0.3791 -1.1118 -0.9370 -0.0580
#>  0.2208 -0.0873 -0.2856  0.5174  0.6270 -0.5069  0.3065  0.1340 -0.6001 -0.2821
#> -0.8515 -1.0797  0.4001  1.6636  1.6421 -0.1078  0.0317 -1.1192 -0.4193  0.4158
#>  0.4519  1.5705 -0.2148 -0.3929  1.9478  0.8430  0.4873  0.3841  0.5562 -0.4775
#> -1.3285 -0.7560 -1.8378 -0.0648 -0.6637  1.3198 -0.9037 -0.4029  1.4633 -1.5583
#> -0.3603  1.2194  0.5936 -0.4421  0.1130  0.5675 -1.2788  1.1951  1.0143  0.4162
#>  0.3284 -0.2767  1.3381 -0.5342  0.1785  0.9531  0.9375  0.1436  0.5063 -1.5100
#>  0.4196  0.9733  0.1567  0.6503 -0.1757 -1.3970  1.7037  0.2430 -0.0224  0.7324
#> -0.6822  0.9205 -0.5434 -1.7000  1.7121  1.1786  1.8219 -0.2921  0.2702  0.7727
#> -0.6492 -0.0276  0.1750  0.0214  0.5516 -0.6291  0.2270  1.2594  0.2236 -0.7209
#>  1.5580 -0.7794  1.0990 -2.6062 -0.2708 -0.2683  0.0037 -0.7121 -1.4032  1.6287
#>  0.9718  0.3636 -0.7233 -2.2607 -1.8592  0.3846 -1.2805 -1.8035  1.0245  0.6583
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]