Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.2924  0.0372  0.4200 -0.4385 -0.2478 -1.6880  0.6878  0.7147  0.3580 -0.6453
#> -0.5177 -0.3338 -1.0734 -0.3699  0.4800 -1.4755  1.3190 -0.6780  1.0366 -1.0246
#> -0.7634  0.1223 -1.7152 -0.6072 -0.0455 -0.9553 -1.0901  0.1822 -2.0714  0.5281
#>  0.8098  0.2979  0.0839  0.0184  0.6431 -0.5577  0.1711  0.4649  0.2711 -0.0205
#> -1.2699  1.2901 -1.5476 -0.0526  0.8675  1.2530  0.6441 -1.9219  0.8156  1.9301
#>  2.0295 -0.0478  0.2612  1.2689  0.3806  0.8796  0.8671  1.1394 -0.1119 -0.7373
#>  0.4740  0.3796 -1.3053 -1.3382 -0.5971 -0.3258  2.5570 -0.1790 -0.2922 -0.2122
#>  1.0330 -0.2436 -0.4407  0.1000  0.0835  1.0258  0.7130  1.0868  0.3454 -0.6191
#>  0.2239  0.4921  1.7315 -0.8135 -0.2091 -1.5003  0.2651 -2.2609  0.2445  0.0123
#>  0.8738  1.0217 -1.0671 -0.4786 -1.0607  0.0148 -1.7206 -0.7029 -1.1292 -0.5008
#>  0.6333 -0.1005  1.9147 -0.1305  0.8976 -0.1441 -0.2884  0.7089 -0.1888  1.7598
#>  0.0738 -0.1544 -0.7972  0.1765  1.1415 -0.0098 -1.0774 -0.4785  0.8742  0.6847
#>  1.9575  0.3312  0.3056  0.1117 -0.3673  1.3167  1.8132  0.1847  1.4817  1.1461
#> -0.3402  0.6002 -2.0182 -2.4500 -0.2171 -1.6330 -1.3086  0.0172  1.3874 -1.3231
#>  0.4345 -0.9128 -1.3688  0.2866 -1.5740  0.5437 -0.1373 -0.7519  0.3186  0.0458
#>  0.3988 -0.7091  0.5305  0.4212 -0.0975 -2.0886 -0.8348 -0.8449  1.0422 -0.8708
#>  0.3402 -1.5230 -1.6258 -0.1907 -1.3209 -0.6565  0.8983  0.0199  0.0016  0.4902
#>  1.2379  0.7236 -1.0774 -1.1850  0.5062 -0.6507 -0.6158 -1.2463  0.4355  1.9296
#>  1.0513  0.2915 -0.9663  2.4198 -0.5025 -1.2458 -0.6250 -0.6750 -0.0362 -0.1086
#> -1.1166 -0.6910 -1.7018 -0.1550  1.5483  0.2047  1.8871 -1.5645  0.2262 -1.6351
#> -0.5783 -0.7249  0.4887 -1.2754  0.2397  1.5501  0.4707 -0.7545  2.9188  0.7484
#> -0.0778  3.3359 -1.0682  1.6365 -0.5918  0.4496 -0.4261  0.0855 -1.6012  0.4412
#>  1.2197 -0.2800  0.3784  0.4634  0.6529 -0.1383  0.5207 -0.3319  0.0085  0.0031
#>  0.6335  0.9732  0.3728  1.0362  0.1936 -2.6375  1.3768 -0.0216 -0.1947 -0.0165
#>  1.6551 -0.1836 -0.0742  1.2379 -0.5407 -0.3596 -0.6562 -0.8387 -0.2516  1.8540
#> -0.6528  0.9956 -0.3140  0.9119  0.1931  0.2396  0.7768 -0.4511 -0.1265  1.1787
#>  1.5249  1.2493 -0.1817  0.0363 -2.2268 -1.0112 -0.8203 -0.0899  1.2813  0.9515
#> -0.7759 -0.6472  1.9775  0.0535  1.0848 -0.9741 -0.2043 -0.1700  0.1464  0.4205
#> -1.9409  1.4710  1.3076  0.7344  0.5031  0.0730 -0.9186  0.0736 -0.0808 -0.7635
#> -1.1287  0.4387 -0.3055 -0.8494  1.8240  0.7102  0.3233  0.1382 -0.0761  2.5932
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]