Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.5349  0.0926  0.3129 -2.0057 -0.1794  1.6552 -0.1212  0.8738  1.3709  0.0412
#>  1.6542  0.7656 -0.3858  0.2159 -0.5117 -0.5081  1.3028 -0.6194 -0.1366  1.3561
#>  0.5614  0.0583 -1.4638 -0.3534  0.4451 -0.9568  1.5808  1.2260 -0.2873  1.7191
#>  0.5133  2.1869 -0.6878  0.2147 -0.3931 -0.0295 -1.2237 -0.7927 -0.7443  1.0086
#> -0.0235  0.3050 -1.1346  0.0199  0.1846  0.8733  1.3861  0.2587 -1.0285 -0.5358
#>  0.5215  0.9289 -0.2849 -0.1618 -0.5927 -0.7485  0.1589  0.8802 -0.0152  0.7132
#>  0.5349  0.9646  0.0199 -1.0323  1.0621 -0.0345 -0.4317  0.7630  1.0758 -1.8694
#> -0.1955  0.5769 -0.6199  1.1204  0.5114  0.1197  1.1403 -1.4189  0.0510  1.0850
#>  1.3000  0.8660 -0.1627 -0.0550  1.1745 -0.0056 -0.8167 -1.1136  1.0375 -0.2905
#>  0.5996  1.1339 -0.0418  0.4983 -1.2893 -0.6593  0.5775 -1.3574  0.9297  0.5162
#> -0.0413 -0.3989  1.4441  2.0016  0.3278  0.5630 -0.9374  1.7414 -1.3988  0.2530
#>  0.8600  0.9165  1.3017  0.7745  2.0770 -1.1349  0.6897 -1.5413 -0.6108  0.9005
#> -1.9846  1.1010  0.5343  2.0603  0.0100 -0.2508  0.3368 -0.5368  1.3483 -0.8991
#>  0.8180  0.1301  0.4697 -0.2054 -0.0920  1.1424 -0.6487 -2.3176 -0.0764  0.2617
#> -0.0315 -2.0286  0.0588 -1.2367  1.5297 -0.0940  1.5485  1.2132 -0.5509  1.2685
#> -0.6675 -0.2515  0.5442  0.1515  1.5227  0.3479  0.2419 -0.5456  0.2675 -0.7455
#> -1.0636  1.4239 -0.1343 -0.0079 -0.9897  0.7765 -0.5585 -1.4315  0.7916 -0.2696
#> -0.6691 -1.2415  0.3217 -1.3274  0.0434 -1.4057  0.4211 -0.4549  0.3220  0.3015
#>  1.3187 -0.6270 -1.6756  0.2681  1.1392  0.2768  0.8649 -0.8892 -0.5926 -0.3823
#>  0.5002 -0.7522 -0.1315  0.0102  0.1900  1.1463  0.5674 -0.1999  0.1602 -0.7536
#>  0.4319  0.9763 -1.1186  0.7413  0.2853  0.1965 -0.6470  0.2248  1.0427  0.7763
#> -1.0521  0.4841 -1.7872  0.2742  1.0555 -1.2245  1.0665 -0.6367 -0.9435 -1.1170
#>  0.2159 -0.0935 -0.1777  1.4057 -0.4199 -0.1419 -1.0467  0.6056  1.7622  0.7451
#>  1.5774  0.3816 -1.4993  0.9976 -0.5610 -0.9150  1.0381 -0.1404  0.3718 -0.7434
#> -2.2958 -1.1582  1.9540 -0.1726  1.0859  1.2414  0.4578 -1.4880 -0.6431  1.3237
#> -0.7030  0.0372  0.7972  0.7051  0.5740  2.2208  0.0879 -0.9996  0.0848 -0.3774
#> -0.0508  0.3288  0.7383 -1.7391  0.1600 -0.0556 -0.8659  0.9708  0.3375 -0.9462
#> -1.3994 -0.8206 -0.2089 -0.0004  0.3130  0.4658  1.2794  2.7617 -1.3821 -1.4721
#> -1.5230 -0.0997 -1.2637  0.5339 -2.0513  1.2549 -0.5846 -0.8001  0.4148 -1.3399
#> -1.5241 -0.0985  0.8333  0.7056  0.9252 -1.0871  1.1732 -0.0050  0.1941 -0.8603
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]