Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.6068  0.0087  0.8906  1.0096 -0.4361 -1.0050  1.0376  0.5152 -1.2808 -0.6536
#>  0.3134  1.4783  0.0998 -0.3030 -0.5019 -0.1775  0.1779 -0.4226 -0.9866 -0.7333
#> -1.2251  0.2134 -0.2086  0.2892 -0.6088  0.7058 -0.1487  0.6877 -2.2582  0.1583
#> -0.2642 -0.7144 -0.9256  0.6682  0.0249 -0.6983 -0.0973  1.9852  0.5120  1.1490
#>  1.0755 -0.5969  0.3015 -0.2489 -0.2739 -1.0615  1.0693 -0.1217 -0.2910 -0.3627
#>  0.9900 -1.6466 -0.7690  2.4066  1.3745 -0.5144 -1.0314  0.1723 -1.9400 -0.0121
#>  1.2322 -1.1624 -0.9474  0.2037 -1.4560 -0.5591  1.1981 -1.8434  1.0323 -0.7817
#> -1.1944  0.4736 -0.2329 -0.6705 -0.9183 -1.1049  0.3146  0.2918  0.7360  0.5126
#>  1.5715 -0.9121  0.1585  0.6741 -0.8753  1.7716  0.1515 -1.6479  0.6884  1.6657
#> -0.2441 -0.0889 -0.4034  0.1426 -1.3056  0.3852 -1.0480  0.1470  2.0921 -1.0523
#>  0.7188  0.9047 -0.0210 -0.1024 -1.3053 -0.4480 -0.0485  0.4754  0.4843 -1.2267
#> -1.1205 -1.3022 -0.0800  1.3837 -1.8282 -0.6061 -0.2066 -1.3884 -0.3252 -0.1669
#> -0.2259 -0.7120 -0.6375  0.7975  0.5198 -1.6618 -1.1408 -1.5013  0.2484 -2.0663
#> -1.1148  0.1976  0.6786  1.7131 -1.6569  1.7253 -0.5174  1.1083  0.6087 -0.0271
#> -0.0689 -0.6399  0.1025  1.1841  0.3723  1.6426 -0.1991 -0.8535 -0.6340 -0.0381
#> -1.3585  1.6424 -0.5318  0.5717  1.5836 -2.8662  0.3958 -1.3428  0.0047 -1.4157
#> -0.7623 -1.4397 -0.6002 -1.6852  0.4494 -0.2136 -1.0878 -0.3331 -1.1281 -1.1591
#> -1.4683  0.0712  1.6625  0.3854  0.8467 -1.0766  0.9176 -1.3963  0.3014  0.2630
#>  0.5698  0.9598  1.5030 -0.2640  1.0881 -0.6649 -0.3967 -0.7662  0.2438  0.5668
#>  0.9082 -0.3108  0.1881  0.2514  0.1172  0.7343  0.0775 -0.7669  0.8924 -0.0205
#> -0.4484 -0.8100 -1.4619  0.3393 -1.7166 -1.5592  0.4916 -0.9596 -0.8674 -0.6002
#> -0.1811  0.2483  0.1683 -0.2707 -0.4230  1.1351 -1.5887  1.6273 -0.6463 -0.3186
#> -0.6148  0.8808 -0.2257 -0.8366 -1.0052 -1.0362  0.2726 -0.3632  0.6031  2.9842
#> -0.9933  0.4475  0.6371  0.6602 -0.1793 -1.0851 -1.2198  1.6202 -0.3279  1.1485
#>  0.5905 -1.6560 -0.0248 -1.2656  1.1602 -1.4906  0.4355  0.9915  0.1642  1.0630
#> -1.3480 -1.3079 -1.3281 -0.5101 -0.8343  0.2808 -1.2696  1.1453 -0.5493 -1.2772
#>  0.0656  1.4021 -0.4469  0.5842  0.6582 -0.6187 -1.0973 -0.1407 -0.0322  0.7968
#>  0.7238 -0.7427 -0.9237  0.3549  0.5909  0.1694 -0.2069  1.6061 -0.4604  0.5252
#> -0.3179  0.8324 -0.9614  0.4087  1.3381 -0.6128 -0.9914 -0.3124  0.4029 -0.0101
#>  0.2701 -0.1351 -0.4685  2.1666 -0.5366 -0.4214  1.1700  2.8752  1.8775 -1.0615
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]