Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6-1.0496e+00 -6.1973e-01  1.4097e+00  8.5459e-01  1.0557e+00  5.1268e-02
#> -8.0125e-01 -5.2686e-01 -2.9616e-01  4.7046e-01  6.9332e-01  2.1248e-01
#> -2.2395e+00  5.9216e-01 -6.9680e-01  6.7014e-01  5.5201e-02 -9.6563e-01
#> -5.3833e-01  1.2326e+00  1.3200e+00 -6.5620e-01 -7.5161e-01  6.8768e-02
#>  1.3383e+00 -9.5800e-01  5.0106e-01 -5.6422e-01  8.6112e-01  2.8945e-01
#>  1.3870e+00 -1.5734e-01 -1.2154e+00  1.4104e+00  6.5156e-01  1.1636e+00
#>  9.9944e-01 -1.2369e-01  1.6091e+00 -2.1259e+00 -4.6035e-01 -2.2216e-01
#> -1.2474e+00  1.7784e+00  2.4159e-01 -1.4116e+00 -4.0945e-01  5.8783e-01
#>  7.0977e-01  1.2064e+00 -1.3741e-01 -1.3747e+00  4.5404e-01 -1.6346e+00
#>  2.2986e+00  6.1981e-01 -1.0685e-01 -1.3197e+00  1.0831e+00 -1.0987e+00
#> -7.0196e-02  1.0009e+00  1.4271e+00  1.1705e+00  1.5817e+00  8.8205e-01
#> -8.5224e-01 -1.2278e-01  2.6451e+00  3.5951e-01  9.1799e-01  7.9501e-01
#>  1.4949e-01  6.6635e-01 -1.8541e+00  1.6938e-01  1.0412e-01  8.2487e-01
#>  1.7861e-02  3.2876e-01 -2.4526e+00 -7.1605e-02 -7.7721e-01  1.2966e-01
#> -7.1374e-01 -1.2840e+00 -1.4127e-01 -2.4846e+00  8.4624e-01  8.6702e-01
#>  2.8935e-01  3.8026e-01  5.9555e-01 -4.8762e-01  1.1938e+00 -1.6192e+00
#> -6.5702e-01  1.0734e+00  1.2630e+00  1.3359e+00  2.4627e-01 -2.8247e-02
#>  5.9623e-01 -4.9010e-02  6.4101e-03 -6.5429e-01  1.6624e+00  1.4045e+00
#>  6.7753e-01  4.5140e-01  1.7018e+00  6.2357e-01  6.0329e-01  5.0864e-01
#>  3.8331e-01  4.7621e-01 -1.8161e-01 -7.8185e-02  9.2485e-01  1.3768e+00
#> -1.4905e+00  1.0892e+00  1.5330e+00 -6.3679e-01 -1.2038e+00  6.5669e-01
#> -1.0469e+00 -4.8677e-01  1.0547e+00  6.2854e-01  1.2720e+00 -4.9608e-02
#> -2.0655e+00  9.1865e-01 -1.2283e+00 -2.0887e+00  8.1409e-01  2.6511e-02
#>  4.6782e-01 -1.1043e-01 -4.5069e-01  2.2510e+00 -1.6884e+00  4.7318e-01
#> -2.9841e-01 -8.6125e-01  1.1040e+00 -1.6975e+00 -1.8245e+00  1.6838e+00
#> -1.0475e+00  6.9954e-01  4.5630e-01 -6.2143e-01  1.9668e+00 -1.2125e+00
#>  1.5588e-01  6.3141e-01 -4.9399e-01 -1.5548e+00  1.2739e-01 -1.9435e-02
#> -1.0426e+00  1.4830e+00 -1.2327e+00  3.3073e-01  2.0251e-01 -1.5040e+00
#> -7.0972e-01  8.2188e-01 -2.7268e+00 -1.0867e+00 -5.6607e-01 -3.0490e-01
#>  6.4391e-01 -8.1897e-01 -1.5334e+00  1.5890e+00 -2.3947e+00 -1.2491e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]