Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.6312 -1.1581 -1.3109  1.1218 -0.1775  0.2469  0.8219 -0.9561  2.3032 -0.2876
#> -1.3247  0.5767  0.2198 -1.0849  1.0035  0.7174  0.0949 -1.0560 -1.4076 -1.4774
#> -1.3419 -0.5579 -0.0618 -0.5633  1.7059 -1.3366  0.1435 -0.8016 -1.2317  0.5577
#> -1.9816 -0.0447  0.9157  0.1484  0.9481  0.3686  1.1302 -0.6646 -1.1496 -0.0247
#>  0.0437 -1.9500  1.7209 -1.0757  0.2388  0.6976 -0.0710  0.6202  0.7582 -0.9148
#> -0.3787  0.9899 -0.4267 -1.8500 -0.7760  1.5431  0.2978 -0.4091 -0.4575 -1.4066
#> -0.1212 -0.1224 -0.3731 -0.5823 -0.6452  0.2681  1.2066  0.7033  2.1370  0.1908
#> -1.1523 -0.4920  0.8365 -0.1491  2.3451  0.8103  0.3746  1.8970  1.7079  0.4379
#>  0.3635 -1.2368  0.4512 -1.0816  0.0285  0.5402  0.1447 -1.0577  1.0949 -0.5021
#>  0.6227 -1.6800 -1.7693 -1.1896 -0.8507  0.8025 -0.4830 -0.1192 -0.6700  0.0012
#> -1.4979  0.5256 -0.4217  0.7005  1.1480  1.2123  0.1101 -0.3708  1.2730  0.1637
#>  1.3234 -1.0936 -0.0286 -0.1599 -0.6950  1.0739  0.6166  1.6449  0.5579 -1.5674
#> -0.4084  0.5496  0.9684  0.2290  0.3778  0.7653  0.1217 -0.6416  2.1285 -0.2262
#> -0.5427  0.2660 -1.6799 -0.0892 -0.6312 -0.1988  0.0351 -0.2128  0.6649 -0.5493
#>  0.2789  2.5185  0.3079 -0.0480  0.8817 -1.5028 -0.3797  0.2056 -0.6117  0.2100
#> -0.8111  0.5401  0.3027  1.8993  0.3359  1.3416  0.6697  0.8249  0.1284  0.1865
#>  1.4444 -0.0388 -1.9248  1.0015 -0.5767 -0.9072  0.0021 -0.0016  0.5456  1.5007
#> -1.2558 -0.1236  0.2803  0.2389  0.5010 -0.0146 -0.1279  0.4303  3.1205 -0.0630
#> -1.0423 -0.0329 -1.1822  0.0664 -0.1452 -0.0571 -0.0544  0.7581 -0.2540 -1.1506
#> -0.5238  1.9484 -0.0798 -0.4349 -0.3074 -1.3513  0.2368 -0.9570 -0.4506  0.3990
#> -2.5276  1.4680 -1.3795 -0.5741 -0.4663  1.8777  0.5923  0.2871  0.7250  0.5471
#>  0.5251  0.0561  0.4528 -1.0142 -0.5982  1.0647  1.1939 -0.1778  0.7771 -0.0540
#> -0.0540 -1.3031 -0.4587 -0.3638 -0.8132 -2.1472  0.9584  0.0876 -1.1595  1.5342
#> -0.1919 -1.0673  1.3120  0.9324  0.7192 -0.1619 -0.2300  0.3781  0.1815 -0.6699
#>  0.2546  0.2400 -0.7757  0.8672  0.1828 -1.2119  0.5058 -1.4115 -1.1456  0.3799
#> -1.0764  1.0036 -3.1558 -0.2686 -0.2670 -1.1266  0.2415 -1.0396 -0.9185 -0.8740
#> -0.3593  0.3162  0.1603 -2.6117 -1.7196 -0.8856  1.4657 -0.9083 -0.7913  0.7457
#>  0.9184 -0.4449  0.4469 -0.1965 -0.6428 -0.7888 -2.5855 -0.9523  0.5756  0.8953
#>  0.3679 -0.7037  0.9372 -1.2634 -0.1744 -0.5657 -0.1793  1.8635  0.1619  1.3965
#>  0.5284  0.9545 -0.3948  1.0695 -3.6077 -2.1564 -0.3839  2.0015  1.0499 -0.0384
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]