Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.3450 -1.0293 -0.1530 -0.4954  1.2745  1.9947  0.1367  0.2079  0.3362 -0.2376
#>  1.5666 -0.4029 -0.5622  0.9261  2.2302  0.0545 -0.6660 -0.7480 -0.5281  0.7146
#> -1.0820 -0.9851 -0.1712 -1.1231 -0.4739 -1.4442 -0.4685 -0.0960  0.8013 -1.0730
#>  1.1500 -0.6884  1.2477 -0.0096  0.2420 -0.4385  1.5504  0.3721 -0.2645  0.2334
#>  2.8501  0.8148 -0.7536 -0.5789  0.8667  0.3630 -0.6015 -1.8356 -2.1797 -1.0578
#> -0.5637 -1.6507 -1.1924 -0.0965 -0.9325 -0.1734  1.5266  0.2189  1.3971  1.0861
#> -1.3945  2.2034  0.3016 -1.6773  0.9927  0.4582 -0.3654  1.2445 -0.1434 -0.6696
#>  0.4339  1.1507 -0.0620 -0.2911  1.1907 -0.1020 -0.5948 -0.8003  0.0213 -0.2746
#> -0.4669 -0.2772 -0.8397  1.2532  0.4342 -0.9944 -0.1503 -1.2716 -2.0330  2.1177
#>  0.0264 -0.2417  1.2665 -0.3464 -0.5028 -0.3381  0.4456  0.4376  2.2474 -0.9221
#>  1.6560 -2.5121  0.3734 -0.5770  0.0043 -1.1592  0.8240 -0.9387 -0.7180  0.8817
#>  0.0652 -1.9501 -2.8335 -0.4365 -1.0484  0.9896  0.6643  1.1243 -0.2113 -0.8912
#>  1.3859 -0.7157 -0.4796  0.8297  1.2429  0.2143  0.5753 -0.3191 -1.0337  1.1615
#> -0.6858 -1.6134 -0.4725 -0.2300  2.5953  2.3929 -1.0382 -1.2425 -1.8250  2.5310
#>  0.0658 -1.3334 -0.4597 -0.6273  1.0843 -0.3601  0.0259  0.4127 -0.4942  1.8911
#>  0.0748 -1.5778 -0.1015  0.4019 -0.3207  1.0295  0.7304  0.9837 -0.6977  0.1103
#>  0.1217  1.3252  0.6239  1.4588  0.4134 -1.8518 -0.3425  0.1403  0.6522 -0.3056
#> -1.9897  0.5831 -0.8815  0.1938  0.6964  0.5837  1.3295  0.5785 -1.1585 -0.1234
#> -0.5076  1.9003 -1.3563  1.8506 -0.9350  0.0481  0.6841  0.7152 -1.3960  0.0388
#>  0.1698 -0.6016  0.5365 -0.1990 -1.2830 -0.3912 -0.0656  0.1833 -0.1192 -0.5728
#> -2.1273  0.4490 -2.2443 -2.0726 -1.5177 -0.2374  0.6044  1.0833  1.7143 -0.2414
#> -0.5775  0.9113  0.4109  0.7338 -0.8816  1.0942 -0.6841  1.1857  0.3931  1.3951
#>  0.8941 -0.0416 -1.0001 -1.3454 -0.1517 -0.4309  0.1861 -0.6194  0.3126  0.1507
#> -0.8099 -0.8422  0.1023  0.0820 -0.7370  0.8865  1.8307 -2.2108 -0.1155  1.0639
#>  0.8340 -0.5948  0.9534  0.7745  0.5291 -0.6444 -1.4436  1.3046  2.0156  0.1888
#> -0.3020 -0.3398 -0.7787  0.6414 -0.1524  0.0265 -0.9287  0.3011  0.0430  0.0489
#>  1.0802 -0.2149  1.0169 -0.4514 -0.7590  0.2275  0.1718 -0.4254 -0.5570 -1.0078
#> -1.8880  0.3739  0.4683 -0.2619  0.2058 -0.2117 -0.0591 -0.9683  1.1585 -1.2097
#> -2.6739 -1.3420 -0.4661  0.8733  0.2643 -0.4428  1.2524 -2.2231  1.3481 -0.9048
#> -1.2929 -0.2069 -1.0023 -2.0887 -1.6369  0.6456 -1.1189 -1.6063 -0.0071  0.3517
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]