Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3705 -0.2383  0.6397 -0.5936 -1.2620 -1.4559  2.1712 -0.2034  1.2763 -1.1522
#> -0.2104  1.2448 -0.2800  0.5369 -0.7785  0.3150 -1.2773  0.7664  0.0533 -0.0618
#> -0.8776  0.7701 -0.6987  0.4971  1.7312  0.5839 -1.0279 -0.8168 -1.1165  0.0872
#> -0.7008  0.8689 -0.9820 -1.4490 -0.7411 -2.0084  1.2672  1.2338 -0.1201 -1.8071
#> -0.3819 -1.2274 -0.5722  0.9501 -0.9748 -2.4928 -1.2900 -0.1041  0.3145  0.1973
#>  0.6262  0.2124  0.0499 -1.0481  0.0052 -1.8065  1.5764  0.2961  1.3466  0.4589
#>  0.9548 -0.9195  0.2789 -0.0132  0.4763 -0.1318  2.3945  2.4525  0.7928 -0.6075
#> -0.5253  0.0371  0.7693 -0.8471 -1.7103  0.2256  0.5777 -0.1777 -0.6065  0.7207
#>  0.9768 -0.5523 -0.7238  0.9376  0.3634 -0.2176 -0.1464  0.2097 -1.3893  1.8589
#>  0.1821  1.6226 -1.4664 -0.9871  1.2283  0.8728  1.3926 -1.1648 -1.8658 -0.1373
#>  0.2022 -0.4279 -1.0912 -0.2500  0.6085 -0.8728  0.7367 -1.1185 -0.2809  1.3310
#> -1.0409 -0.4474 -0.0666  0.1663 -0.9325 -0.2513 -0.2250  0.0365 -0.3632 -0.0733
#>  0.4319  0.7238 -0.1026  0.6506  2.3886 -0.6595  0.2872 -1.2091  0.4629  0.2094
#>  0.1887 -0.4099  1.8896 -0.6634  1.8979 -0.9707  0.4241  0.2149 -0.0414  1.5391
#> -0.5475 -0.6363 -0.1163 -0.0003  2.8236  0.8648  0.0309  1.9718  1.4789  1.5210
#> -0.1799  0.7371 -1.1358 -0.7915 -0.5905  1.1682  0.8320 -0.0770 -0.8368  1.4348
#> -0.0574  0.3994  0.4065 -2.3473  0.0696 -0.3133 -0.6284  1.5441 -0.4274  2.2386
#>  1.7165 -0.2823 -0.2442  0.2345  0.8582 -1.1547 -1.1762 -0.2581 -0.0147 -1.7678
#>  0.4714 -0.8923  1.4521 -0.4860  1.0739 -0.8540  1.0410  0.2612  0.7431  0.0293
#> -0.1590 -0.8166  0.0936 -0.1130 -1.6231  0.5185 -1.1718 -0.7283  0.1845 -0.3592
#>  0.7796  0.8306 -0.6106  0.4407 -1.9405  0.2750  1.5078  0.7583  0.4364  1.8531
#>  1.8729 -0.2570 -0.1417  0.3589  0.5100 -0.0947 -0.4787 -1.6045 -0.3026 -0.9410
#> -0.6934  0.1629 -0.7882  1.2078  0.3263 -0.5851 -0.1435  0.1326 -0.2889  0.0828
#> -1.6832 -0.5380  1.3554 -0.7399 -0.6358  0.1413  1.0579  0.4663  0.1527 -0.4426
#> -1.1722  0.5689  1.7715 -0.5510  0.7538 -1.1101 -0.3974  0.8793  0.2306  0.2211
#>  0.0323  0.4027  0.8646 -0.3484  0.7942 -0.2639  0.0570  0.1805 -0.2193  0.0068
#>  0.8768  0.5503 -0.3096  0.6738 -0.0972  0.2099  1.8998 -0.4527  1.6886  2.6766
#> -0.0997 -1.2606 -0.6860  0.6405 -1.2759 -1.2462 -0.3627 -0.6345  0.9097 -1.3780
#>  0.0186 -0.4654  0.5789  1.8035 -1.8816  0.9929  1.0757  0.6670  0.5329  1.4094
#> -0.5930 -0.8111 -0.5041 -0.2746 -0.4107 -1.1909  0.1205  0.0926 -0.0073  0.3066
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]