Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2993 -0.7285 -0.4953  0.0098  0.7479  1.4589 -1.2543  0.1938 -0.6943  1.0706
#>  0.5340 -0.6802  0.9154  1.6391 -0.5645 -0.2556  1.7396  1.3791 -1.0325  0.6323
#>  1.2597  1.9301 -1.9940  1.3620 -1.0970  0.0047  0.6758 -0.2155  0.1487  1.2568
#> -0.6601  0.1082  0.2547  0.1354 -0.7790  1.0087 -0.3748 -0.1414  1.2450 -1.2279
#> -0.3384  1.0212  0.4599 -0.3868 -0.4117 -1.6064 -0.7705  1.8346  0.6195 -0.7144
#> -0.3823  0.3087 -0.6741 -0.1217  0.2042  1.7061  0.0925  1.2551  1.1854  0.6132
#>  1.1581  1.1711  1.5860 -1.2076 -0.1436 -0.4288  0.9509  0.7354  0.4880 -0.6380
#>  0.8227 -1.4864  0.0712 -0.6332  0.4378 -2.3053  0.5836 -0.0723  1.6079 -0.3564
#>  1.5389 -0.6072 -0.5087  1.5678 -0.1879 -1.5803 -0.4772  1.6788 -0.0579  0.8727
#>  0.1899 -1.1844  1.5427 -0.1602  0.6597 -0.3636  0.3303  0.6243 -1.1417 -0.3286
#> -0.1137  0.3897 -1.4310  0.0792 -0.3115 -0.9834 -0.0102  0.4403  0.0904  1.2226
#> -2.2068  0.4279  1.0270  0.4653  1.8377  2.0562 -1.4801 -0.8778 -0.3064 -0.1525
#>  0.3865 -1.2724 -0.3512  1.7205  0.4147 -1.4044 -2.4678 -0.3525 -0.7320  1.2291
#>  0.7918  1.0502  0.9918  0.2864  0.2262 -0.4483  1.6047 -1.8478  0.4643  0.5688
#> -0.7093 -0.1070  0.2964 -0.1445  0.2082 -0.8745  0.5918 -0.3761  0.7586 -1.7501
#> -1.9535 -0.3348  0.5297 -0.9237  0.4683  0.4162 -0.1001  0.2621  1.3750 -2.2599
#> -0.7546 -0.7496  0.8940 -0.1642 -2.1360  0.9203 -0.1181  1.3530 -0.4513 -1.8992
#> -1.6527  0.7596 -0.1632  0.4211 -0.0748 -0.8817 -0.0336 -1.2715 -0.0298  0.3782
#>  0.8366 -0.4766 -0.0522  0.1227  0.5848 -0.6887  0.8011 -0.0632  1.1412 -0.3865
#>  0.4009 -0.0857 -0.2888 -0.0074 -1.8719  0.0938 -0.1729  0.6847 -1.7524  0.5096
#>  0.1170  0.5042 -1.2444 -0.9812  0.6180 -2.3802 -1.1835  0.8595 -1.3591 -0.7310
#> -0.0290 -1.1839  0.4272 -0.3940 -0.6436 -0.3207  1.0478 -0.0108  2.2967  1.3091
#>  1.0974  0.8074 -0.2951  0.4331  1.2711 -1.1046 -0.2303  0.3296  0.6401 -1.2227
#> -0.4642  1.1821 -1.0628  0.0552  1.1970  0.5486 -0.4816 -0.0080 -0.2049  0.8680
#>  1.1866  0.5233 -0.7968 -1.6183 -1.5449  0.5855 -0.0030  1.0797 -1.5502  1.0665
#> -0.5841 -0.3304  2.0924  0.9954 -0.5409  1.1811 -0.4202 -1.1144 -0.2390  1.8055
#> -0.9443 -1.0045  0.1144 -0.1519  1.3511 -0.7249 -2.0815 -0.4716 -0.1511 -1.8638
#>  1.3313  1.2222 -0.3332  1.5946 -1.3419 -0.1803 -0.1510 -1.5398 -0.8243 -1.6159
#>  2.5560 -1.2001  1.2044  0.6515 -0.4139 -0.3935 -0.5128 -0.3911  0.3158 -1.6278
#> -1.4237  0.8071  1.7048  0.2536  0.0676  0.4384 -0.8898 -0.0941 -0.1981 -1.0515
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]