Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.4263  1.3232 -0.4303 -0.5798 -1.7652 -2.3771 -0.2705  1.5455 -0.3175 -0.8000
#>  0.0520 -0.0122  0.4653 -1.7652  1.2120  1.1223 -1.0252 -1.4448 -1.2818  2.7787
#>  0.7116 -0.5182  0.9076 -1.0817  0.7095 -0.8623 -0.1028  1.9915  1.7302 -1.0282
#>  0.1093  0.2907 -0.8910  0.4874  2.1302  1.4369  1.5923 -0.4403 -0.4998 -2.3265
#> -0.8189 -1.7366 -0.9379  0.9468 -0.1608 -0.5758 -0.8023  0.8535  1.6115 -0.2375
#>  0.2010  0.5198  0.9157 -1.0258  0.2557  1.0022  1.2999 -1.3766 -0.3032 -0.3168
#> -1.2330  0.0551  2.9958 -0.5747  0.8948 -0.5499 -0.9289 -0.0199 -0.9845  0.2512
#> -1.4709  0.1206  0.7902  0.9921  0.0042  0.5260 -0.2795 -0.1115  1.3487 -0.1389
#> -3.2253  0.8810  0.8469  2.3678  1.3215  0.1105 -0.1971 -0.9863  0.6966  0.5771
#>  0.0982  0.8603  0.9660  0.3112 -0.9524 -0.9529 -0.3167 -0.7363 -0.7456 -1.1185
#>  0.7328  1.7378 -0.4409  0.0748  0.3225  0.4174  0.7399  2.1791 -1.1467 -0.9786
#> -1.0159  0.6746 -0.1073  0.3401 -0.5465  0.7956  0.2467  1.3509 -0.3908 -0.4884
#> -0.5834  1.0269 -0.8210 -0.9681 -0.1010  1.2753  0.2179  0.9016 -0.2017 -0.6208
#> -0.2437  0.2491  0.0569  0.5642 -0.4239  0.1327  0.4008  0.1406  0.0449  1.3347
#> -0.5149  1.2354  0.2205 -0.2308  0.4119 -0.3604 -0.0247  0.4061  0.6010 -1.5420
#> -0.0696 -0.6667 -0.0550  2.4075 -1.2765 -0.7959 -0.9922 -0.5055 -1.2032  0.1757
#>  0.8655  0.2773  0.5657  0.5346  1.0917 -0.5561 -1.3944  0.6450  1.1119  0.5700
#> -2.1249 -0.3405 -1.5963  0.2832 -0.5764  0.0751  0.3522 -0.4563  0.1499 -1.7171
#> -0.2442 -0.9963  2.0906  0.2303 -0.8753 -0.1070  0.2806 -0.7756  0.3797  0.0042
#> -0.2573 -0.0707  0.9722 -0.3224 -0.4922  0.1478  0.7278 -1.4961  2.1224  0.2456
#> -0.5794 -0.9937  0.1570 -0.9746 -1.4854  2.0529  0.6229 -2.3006 -0.8099 -1.6269
#>  0.1736 -0.6643 -0.2740  1.9250 -0.9836  0.1162 -2.4981  1.1407  0.1838 -0.9648
#> -0.5431  0.4182 -0.7630  0.8151 -0.9704 -0.4991  0.5259  0.3943 -1.2103  1.7347
#>  0.9454  0.2153  1.4272  1.9225 -0.8063 -1.2179 -0.7244  1.4468 -2.9651 -0.7171
#> -0.6072 -1.5360 -0.4618 -0.6539  2.4465  0.8501 -0.5170 -0.3103 -1.9358 -1.2450
#>  0.0951 -0.1205  0.3029  1.4345 -0.2600  1.0477  1.0893 -1.8473  0.7219  1.6665
#>  0.8114  0.8871  1.7903  1.1900 -0.1584  0.0773 -2.4737  1.0873  0.3009  0.8643
#> -0.1383  0.6682 -0.6088 -0.9837  1.0232 -0.0401  1.9629 -0.3319 -0.2160  1.6769
#> -1.0746 -0.1446  0.9742  0.1688 -0.5828  0.5783 -2.0980 -0.5331  0.9455 -0.7538
#> -1.6277  0.3316  0.9000  0.6200 -0.6455  0.2055 -1.0713 -0.0442  0.5037 -0.4589
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]