Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.1595  0.9174 -0.5315 -0.2837 -1.4947  1.1098 -1.1126 -2.9457  1.6178 -1.2337
#>  0.9417 -0.0926 -1.4281  0.2311 -0.6835  0.4079 -0.8288 -1.5048 -0.4314 -1.1660
#> -1.1664 -0.5275 -0.2526  0.7129 -0.8379 -0.1170  0.9273 -0.2235 -0.2015  0.5007
#>  2.6555 -0.7397  0.6444  0.0155  1.8200 -1.1460  1.9496 -1.4594  0.5959 -0.4851
#>  1.1713  0.6321  0.6916 -1.1633  0.6532 -0.1135  0.7193 -0.2160  0.1442  0.1515
#>  0.0346 -1.2055  0.3392  0.2290  1.6897  0.5354 -0.7015 -0.0596  0.0158 -0.8520
#> -0.1741  0.5366  0.7254  1.0081  0.5478  1.8681 -0.1744  0.0416  0.9308 -0.0616
#>  0.0321 -1.1590 -0.4107  0.1295 -0.2132 -0.8223 -2.4582 -0.0710 -0.3969  0.2347
#>  0.3906 -0.6855  1.1955  0.8225 -1.9784 -0.1397  1.3026 -0.7824 -1.1757  0.0957
#> -0.6946 -1.4054 -0.0657  0.1235  0.8701 -0.5622 -0.4496  0.1245 -0.7191  0.5234
#> -0.8874 -1.3468  1.0638 -0.2932  2.5120  0.4884  0.8648 -0.4456  2.5249 -0.1693
#>  1.4670 -1.1727 -2.7438 -0.4053 -0.8422 -1.3341  1.6227  0.6642  1.4095  0.7097
#> -0.2111 -0.2821 -1.4511 -0.3604  1.3008  0.5254 -0.0520 -1.4375 -0.3165  0.1964
#> -0.0648 -0.4266 -0.0228 -0.1994 -0.0251 -0.8244 -0.8539  1.2867  0.7656  0.7070
#>  0.6915 -2.3224 -0.5836  0.2179  1.1789  2.5972 -0.3851 -1.4646 -0.6763 -1.0723
#> -2.1934 -1.1500 -0.2312 -1.3441  1.7420  0.0358 -0.7370  0.3199 -1.2501  0.6021
#>  0.0263 -0.5290  1.3519 -0.2972  0.9786  1.2159  0.5469 -0.6476  0.0242 -1.8793
#>  0.3001 -1.0231 -1.7224  0.2015  1.2649  3.8765 -1.7605  0.7971  0.6632  0.6130
#> -0.0468 -0.9108 -1.5407  0.2591 -0.4133 -0.3196 -0.4448 -1.0201 -0.1376 -1.7216
#>  0.0542 -0.1402 -1.8017 -1.9189  0.3281  0.1858  1.0636  0.0777 -1.9734  0.1426
#>  0.0164 -0.5906 -1.2325 -0.7319 -0.7445  1.2176 -0.1818  0.4790 -2.0980  0.1143
#>  0.4659  0.1453 -2.0708 -0.9943 -3.4994 -0.5818 -0.9569  0.3518  0.2225 -2.1072
#> -0.3435 -1.3815  1.3405 -0.2898 -0.0063  0.4405  0.4414  0.4717 -0.2136  1.6702
#>  1.0628  0.1735  0.9220 -0.6093 -0.4209 -0.1379 -0.1617  0.0074  0.9776 -0.7315
#> -0.7052 -0.5570  0.8167 -0.7880  0.0210  1.7119  1.6062  0.5252 -0.6108 -0.6300
#>  0.6976 -0.3276 -0.8358  0.0509  0.1781 -0.1799  1.6226 -1.5479 -0.5057  0.0414
#>  0.1520  0.4408 -0.5850 -0.7115 -0.9699  0.8202  1.0279 -1.0563 -0.5274 -0.1606
#>  1.4933 -0.2996  2.9584 -1.4142  1.2294  0.8469  0.1298  0.7688  0.5810 -0.5329
#> -0.2546  1.4254 -0.2552  0.3486  0.3689 -0.6820 -1.3199  0.1790 -0.3240  1.1754
#>  0.8239  0.7094  1.1414  0.5920 -1.0247 -2.0251  0.2824  0.9785 -0.9426 -1.1990
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]