Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6228  0.2131 -0.4059 -0.2998 -0.1963  1.2173 -1.2514  1.7746 -2.3242  0.6400
#>  0.9720  0.0407  0.2017  0.1848 -0.5655 -0.1884 -0.3478  0.3185  0.3413  0.0200
#>  0.1411  0.6656 -1.2771 -1.0664  0.1203 -1.0035 -0.1792 -0.8821  0.7208  0.0430
#>  0.3177  1.4430 -1.7706 -0.3739  2.0058  0.5018 -1.5532 -1.2047  0.5488 -0.9321
#>  0.5104 -1.8934  1.0000  0.0125  0.5747  0.0429 -0.8690  0.6091  0.4178 -0.0355
#> -2.9603  0.9207 -0.3462 -1.8441  0.1556  1.4634  0.9416  0.2366  0.2679 -0.4365
#>  0.0096  1.1448  0.3697 -0.8402 -0.1094  2.7381  1.4272 -0.5650  0.4189 -0.2434
#> -0.2106 -0.2514  1.1465  1.9848 -1.9530 -0.2577 -0.0876 -2.8541  1.8454 -1.6368
#>  0.0505 -0.0590  0.7715 -0.9024  1.2567  1.2984 -0.2050  0.1580 -0.9780 -0.0618
#> -0.9790  1.9850 -1.5060 -1.5907 -1.6887  0.7858  0.3422  0.5591 -1.0742 -1.2736
#> -0.3060  0.0451 -0.4851 -0.1511 -1.4369  0.5991  1.3317 -0.7609  1.1795 -0.6201
#> -0.2811  0.6003 -0.8748  0.4874 -0.7523 -1.8091 -0.3771  0.8918 -0.0174 -0.1201
#> -0.8805 -1.3330  0.4748 -1.2822  0.8941 -1.3334  1.0229 -0.4786  0.6878  0.1473
#>  1.6180 -0.2277 -0.2211 -0.0378 -0.2957 -0.0746 -0.8246 -0.8522  1.9824 -1.9512
#>  0.9454 -1.1764  1.6008  0.0433 -0.2263  1.1248  0.1693  0.4078 -1.1789  1.0481
#> -1.1121  0.0784  1.7524  0.0090 -0.4476 -0.5315  0.5705  1.0647  0.5927 -2.3193
#> -1.0024 -0.0293  0.8217 -1.5464  1.6026  0.4357 -0.8175  0.5477  0.5801 -0.8737
#>  0.4378 -1.8953  0.9550 -0.6686 -1.9079  2.3827  0.7183  0.6000 -1.3365  0.6459
#> -0.6091 -0.4977 -2.1638 -0.5817  2.5023  0.0256  0.3400 -1.0170  0.1604  0.6365
#>  0.3394  0.5304  1.2396 -0.6513  0.8382  1.2414  0.6794  2.4127  1.4485  1.9615
#> -0.5166 -1.7180 -1.3775  0.5649  0.1737 -1.7421  0.5038  0.5520 -1.3293  0.6417
#>  1.3035 -0.2160  0.3012  1.7249 -0.5572 -0.1449 -1.0325  1.7400  0.6577 -0.1002
#>  0.2576  1.4399 -0.8621 -1.0326 -0.4188 -1.0746  1.0883  0.2865  0.4292 -0.5054
#> -1.5227  0.4335 -0.8922  0.5120  0.0570 -0.8608  0.4188  0.2340  0.3136 -1.4905
#> -0.5922  0.9123  1.4364  1.5789  0.2391 -1.3231 -0.6913  2.4830 -1.2519 -0.5694
#> -0.0835 -2.1436 -0.8676 -1.5029  0.4647  1.9781  1.0390  1.8703  0.1225  1.0923
#>  0.5081 -0.3543 -0.8979  0.0793 -0.1331 -0.6338  0.8014 -1.4365  0.1601 -1.0871
#>  0.3873  1.1138  1.1521  0.1259 -1.8076  2.1634  0.4195 -0.6008  0.2625 -0.5112
#> -0.4976  0.2953  0.8355 -1.2687  0.7995 -0.2219 -0.0495 -0.9060  0.9297 -1.9398
#> -0.7560  1.9392  0.0050 -0.2470  0.1386  1.5203  0.9031  0.4905  0.7179 -0.7528
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]