Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.5208 -0.3841 -0.6117  0.1615  0.3351  0.4796 -0.2321  0.5956 -0.2118  2.8164
#> -0.5689  0.3030  1.7770 -0.8981  1.7251 -1.8059  0.9448 -2.3975  2.0007  0.9362
#>  1.8782  0.8988 -1.8239 -1.2721  1.1880 -0.2747  0.1426 -1.1311  1.2922 -0.1301
#> -0.8986  0.0886  1.1043  0.9830 -1.7067 -0.0255  2.2114  0.0217  0.0981 -0.0089
#>  0.2238  0.3311  0.1685 -0.9151  0.7239 -0.2758  0.9008  0.1735  0.1507  0.3240
#>  0.2302  0.7506 -0.9328  0.2415  1.3823  0.6242 -0.8045  0.1006 -0.7778 -0.4378
#>  0.7455  0.1767  1.5087 -0.3624 -1.1376  0.3731 -1.0671 -0.5570 -0.2884  1.6826
#> -1.3668 -1.4273  0.4633 -0.7998  0.9064 -0.3516 -0.7546 -2.4051  0.1749  0.9569
#> -0.9958 -0.6455  0.1256  0.6661  1.2932 -0.9400 -0.3891 -0.3049  0.6096 -2.7407
#>  0.5888 -0.3609 -0.1283 -0.6413  1.9570  0.1682 -1.1214  0.8261  0.1715  0.4419
#>  0.4528  1.0584  0.5487  0.0428  0.6833  0.4232 -0.0377 -0.5392  0.6663 -0.3814
#>  0.7843 -0.8130 -0.1885  0.4283 -1.2449 -1.0025 -0.4754 -1.0058 -0.6094 -0.4374
#>  0.8390  1.0851 -1.7313  0.1368 -1.1058 -0.6089  1.3302 -0.0376 -0.4544 -1.4810
#>  0.1813 -0.2619 -0.4542  0.4392 -0.1843  1.2326 -2.1014  0.0602 -0.7393 -0.1696
#> -0.0109 -1.4820 -1.4059  0.9219 -0.7779 -0.0933 -1.4965  1.3480 -0.4141 -0.8231
#> -0.8549 -0.9423 -1.2895 -0.5530  1.0900 -1.9981 -0.4735 -0.9055  0.5055 -0.0831
#>  0.2523  0.7403 -0.6029 -0.5344  0.9344  1.7160  0.1234 -0.1410  2.8575 -0.0231
#> -0.0688  0.0606  0.0960  1.9999 -0.2779 -1.5153  0.8926  0.4075  1.1957 -1.8074
#>  0.8244 -0.8315 -0.6113  1.7217 -1.1817  0.0611  1.0422 -0.1346 -0.8562 -1.0189
#>  0.4136  1.6743  0.0569  2.9461  0.8884  0.5238  1.0676 -0.0967 -1.5119  0.1211
#>  1.3258 -1.0185 -0.9431  0.7015 -1.1344 -0.1761 -1.0860 -0.0189  1.3549  0.2073
#> -0.5433 -1.1498 -0.7745 -0.3839  1.7162 -1.2598 -0.3435  0.2468 -0.4766 -0.5646
#> -1.2742 -0.9043 -0.7191  0.1638  0.0588 -0.5961 -0.8489  0.1282  0.1805  1.9261
#> -0.0398  0.2828  2.0279 -1.3742  0.0717 -0.6799 -0.3611 -0.5110 -0.7913  0.2611
#>  1.3108  2.1100 -1.2263 -0.4164  0.1797 -0.6765 -0.5377  2.0652  0.2229  1.0222
#>  0.4511 -0.8082 -1.0421 -1.3621  0.4048  1.0936 -1.1757 -0.4842 -0.6423 -0.1903
#>  1.0953 -0.8615 -0.4771  2.3787 -0.5111  1.0668  0.4938 -0.0198  0.1575  0.5636
#>  1.2428  0.4551 -0.7470 -1.0165  1.1155 -0.5486 -0.3989 -0.1677  0.9354 -0.4280
#> -0.9499 -1.0848  0.3112 -0.4967  1.0177 -2.4422 -0.2726 -1.1358 -1.6351  1.2080
#>  2.1884  1.6610  0.6608  2.2340 -1.1942 -0.1779  0.2645  0.8904  0.0674 -0.0223
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]