Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.2049 -0.0350  0.5377 -0.2711 -0.4215 -0.6647 -0.5392 -0.6040  1.4372 -1.2323
#> -1.1448  1.0732  0.6261  0.8050  0.7001  0.5384  0.0364  0.1329 -0.1536 -1.0691
#> -1.0384  1.2905 -0.8438  0.5736  1.1730 -0.1676  0.9244  1.7047  0.3543  0.4528
#> -0.3217 -1.9244  0.0105 -1.3314 -0.5793  0.9347 -0.9235 -1.4080  0.7550  0.6402
#>  1.5305 -1.5239  0.0834  0.9024 -0.5653  0.6732  1.0636  0.8167  0.0635 -1.1613
#> -0.1865  0.4688  0.0252 -0.3817  1.8828 -0.7484 -1.9938 -1.6376 -1.6687  1.4032
#>  1.3055  0.1876  0.0700 -0.4351 -0.9343 -0.6821  1.4813 -0.9347  1.7007 -1.1208
#>  0.3992 -0.5092 -0.3649 -0.7191 -1.0588  0.1770 -0.3562  0.5491 -0.5504  0.5458
#> -0.7832  0.4803  0.2882  1.0825 -1.3461 -0.5550  0.1918 -0.4684 -0.6807 -0.2922
#>  1.3228 -0.6543 -2.2420  1.7345  0.1760 -2.0376 -1.0236  1.5964  0.7246  0.5938
#>  0.1875  1.9585  1.0274  0.6276 -0.1667 -0.4312 -0.5523  0.5289 -0.2860  0.1813
#>  1.0446  0.6555  2.0770 -0.5533  0.3799  0.0535 -0.6899  2.6037 -0.8522  0.2801
#> -0.1740 -0.8296 -0.1770 -0.4124  0.1185  1.1985  0.5444  0.4325 -1.1778  0.6357
#>  0.9942 -2.0926 -0.7133 -0.2980  0.5061 -0.2477 -1.2332  0.0551 -0.8861 -1.2146
#> -1.1398 -0.4104  1.4618  1.2365 -0.4214  2.2506 -0.9364  0.0872  0.0859 -1.4603
#>  0.9975 -1.3375 -0.3307  0.8590  0.3626 -0.9632  0.1627 -0.0864  1.1721  0.3859
#>  0.5397 -0.7131 -1.0926 -0.0276  1.5063 -1.1770 -0.7242 -0.6085 -2.4002 -0.5043
#>  0.5771  2.0443  0.7423  1.0321 -0.6682 -0.3587  1.2248  1.0031  0.3067  0.1546
#> -0.2982  1.3332  2.1847 -0.2051 -0.1722  0.9277  1.2266 -0.5253  0.9939  1.0401
#> -0.7422 -1.2357 -0.0082 -0.0300 -0.1794  0.6174 -0.4719  0.9025 -0.6710  0.2734
#>  0.5255 -0.9291  0.7615 -1.4567  0.4159  0.3794  2.0331 -0.5219 -0.0184 -1.6974
#> -1.4149 -0.3514  2.3083 -0.3356 -0.3604 -0.5328  1.4755 -1.4534  0.2038  0.2960
#>  1.1716  0.4688  2.1651 -1.5000 -0.6576  1.0115  1.6938  1.2138  0.4112  0.4777
#> -0.0565 -0.4566  0.9415 -0.3193 -0.2121  0.2147  1.0909 -0.2846 -0.5404  0.1344
#> -0.2620 -1.2838 -1.4442 -0.3395 -0.8563 -0.1039  0.3198  0.0966 -0.3082  1.6661
#>  0.8026  0.4504 -0.3830 -1.0642  0.3156  0.3750  0.2195 -1.4630  0.1289  0.8982
#> -0.0109 -0.5427 -1.8008  1.1758  0.0658  0.9658 -1.6297 -0.3060  0.5306 -0.2204
#>  0.2112 -0.8206 -1.0363  0.7308 -0.2972  1.0568  1.4786  0.6798 -0.3301 -1.1391
#>  0.1558 -0.8530  0.1952 -1.0011  0.0617  0.8744  0.5035  0.0571 -1.0011  0.3598
#> -0.2008 -1.4345  0.1839 -0.6104 -0.3445  0.8983  0.0535 -0.4872  0.7833 -0.5292
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]