Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.8501  1.0040  0.2118 -0.1037 -0.6450  1.5966  0.8611  1.5726  0.4398 -1.0011
#> -0.0860  1.3288  1.2873  0.9877  0.3402 -1.1900  1.9040  1.7755 -0.0603  0.6444
#> -0.8103  0.0643 -1.5921 -1.5574  0.3037 -0.2776 -0.9443 -1.0348 -0.9169 -1.0969
#>  1.9312 -0.3630  1.4260  0.0850 -0.8348  1.1804  0.9496 -0.9382 -0.5292  0.2387
#> -0.4369  1.2265 -0.3958 -1.1908 -0.3798 -0.5710 -0.4465  1.2029  2.4741  0.0179
#> -0.7215 -0.9055 -0.5215  0.6799  0.6800 -0.0851 -0.4531 -0.7685 -0.1425  0.0875
#>  0.7789 -0.3661  0.4253  1.4761  1.4514  1.1160  0.6098  0.0902  0.5496  1.3284
#>  0.2827 -0.0725  0.6466 -0.2973 -0.1905  0.9298  1.8448 -1.3186 -0.1886  0.8151
#>  1.5665 -0.1689 -1.0737 -1.4613 -1.3711 -0.4226 -1.5800 -0.7607  0.0315 -0.1251
#> -1.3633 -0.2475  0.3887 -0.0223  0.3499  0.2048 -1.1354 -0.7347 -1.5395  1.3895
#> -0.0841  0.1347 -0.1205  0.0168  0.5773  0.0826  0.5912 -0.5397  0.9136  0.0547
#> -0.8906 -0.1081 -0.0058 -0.7817  0.9610  0.7473  1.7508  0.0613 -1.7376 -1.7513
#>  1.1050 -0.9060 -0.5605 -0.5741  0.5854  1.1450 -0.2789  1.5014  2.0556 -0.7926
#>  0.0623 -0.0956  0.8631 -0.5113 -0.2715  1.9422  1.4223  0.4260  0.2905 -0.8411
#>  0.4327  0.6238  1.4702 -0.1077 -0.1080  0.5151  0.7700 -2.5596 -0.3743 -0.4249
#>  0.0657 -0.8638  0.9842 -0.0520  0.6485  0.7227  0.7017  0.7256 -0.4672  0.8744
#>  0.5454  1.7795  0.3844  1.5285 -2.2084  0.1449  1.5635  0.4406 -0.0887  0.6957
#>  0.0033  1.2176  1.2239 -2.1100  0.3054 -1.5340 -1.8087  1.2302  0.3420  0.3278
#> -1.8829 -1.4076  0.7004  0.4196 -1.1191  1.5433 -0.3176 -0.7655 -0.1118 -1.7935
#>  1.5207  0.9575 -0.2551  1.0351 -0.7990 -1.0646 -0.0425 -0.9692  0.3465  3.2397
#>  1.9577  0.9094 -0.0045 -0.5708  1.8801 -0.7599  0.6816  1.2719 -1.8779  0.3351
#>  1.2125 -1.1227  1.8637  0.0563  0.3103 -1.6768 -0.0019 -1.1720 -0.3868 -0.1629
#>  1.3571  0.4244 -1.2959 -0.9960 -0.3804  0.0798 -0.0778  0.4349  0.7587  0.5737
#> -1.1392  0.0330  0.4369 -1.0090 -1.5079 -0.0924 -1.0302  1.4330  0.9239 -0.8907
#> -0.6094 -1.3685  0.0091 -1.4661  0.4829  0.1679 -0.5917  0.7528 -0.6674 -0.3455
#>  1.5489 -0.7490  1.3914 -0.4694  0.1536  1.1873 -0.0836  0.0770 -0.9351  1.3762
#>  0.3711  0.2688  0.5495 -0.7670  0.8265  0.5912  1.1389 -0.9809  0.1472 -1.7298
#>  0.0242  0.6279  0.0140  0.8102 -0.2237  0.4439  0.5173 -0.1886  0.6329  0.2728
#>  0.8557 -0.8966 -0.9081 -0.1522  0.0772  0.0779  1.2694 -0.0068  0.8689  0.9244
#> -0.4181  1.4669  0.9990  0.5984 -1.1461  0.8859  1.6291 -0.7719  0.4868 -0.9780
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]