Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0364  0.1256  0.9001  0.5771  0.5855  1.0867  0.3434 -0.2568 -0.3391  0.9002
#> -1.3009  2.4117  1.3242 -0.6657  0.8479  1.1168 -1.1408  0.0246  1.0241 -1.4740
#>  0.4084  0.4046 -0.5385  1.3191  0.0702  0.5556 -0.1631  0.5944  0.2616 -0.6922
#> -0.4100 -1.4797  0.3239 -0.2847 -2.3133 -0.4202  0.7990 -0.0868 -0.4477 -0.4386
#> -0.2969 -2.0722  0.3154  0.3084 -0.2039 -1.2722  0.7054  0.2438 -0.6791  0.5194
#>  0.0101 -0.0113 -0.5881  0.5746  1.0716  0.6997 -1.2506  1.6343 -0.3819  0.3030
#> -1.1791 -0.0640 -0.8982  1.4612 -1.5419  0.1287  0.3104 -0.4497  0.5443 -0.4354
#> -0.7336  0.5682 -1.4141  0.7529  0.8411 -0.6820  0.1050  0.4363 -0.9584  0.9568
#> -0.7615 -1.2031  0.0078  0.0440  0.1741 -2.2446  0.9760 -0.3321  0.4068  1.5624
#>  0.5378 -0.2449 -0.4164 -0.1811  1.9046 -1.4660  0.0215 -0.1421 -0.2898 -0.6399
#>  0.6542 -0.5335  0.4315 -2.0099  0.5767 -0.3171  0.1313  0.9816 -0.1235 -0.1823
#> -1.7111 -0.4333  0.2726  0.1823  0.4901  0.2094  0.0914  0.2739  0.6637 -0.9497
#> -0.3371  0.4861 -1.7216 -0.6587  1.3804  0.4863 -0.7328  0.0646  1.5253  0.0546
#> -1.0090  1.4843 -0.9766 -1.7142 -0.9474 -0.9711  0.6129 -0.0043  0.5742 -0.2691
#>  0.0727 -0.4081  0.3925 -1.2038  0.4633 -0.3544  0.5738 -0.8274 -0.1247  0.8756
#> -0.4660 -0.3306  0.9673 -1.6029 -3.0063 -2.0461 -0.1124  1.0067  0.1589 -0.2993
#> -0.7731  0.0026  0.0189  0.6983  0.5572  0.9403  0.0200  1.1376  0.4587  0.0107
#> -0.1788 -0.1434 -1.4665  0.1171  0.5597  0.1943 -1.3553 -0.7706 -0.7005  1.1535
#> -0.0986 -1.7966 -1.2778  0.0224  0.3936  1.6355 -0.0543 -0.1665  0.1439 -0.0468
#> -0.4482 -2.2124 -0.4134  0.7439  0.6430 -0.7041 -2.3017  0.9467 -1.1197 -0.2464
#>  1.3064  0.3443  0.4433  0.2070 -0.4836 -0.5182  1.0687 -0.1942  1.2026 -0.9734
#> -0.3134 -1.5715  0.4876  1.2193 -1.5275 -0.6083 -0.0342  1.0868  0.2067  1.6599
#> -0.2421 -0.7442 -0.0563 -0.0414  1.0539 -1.1638 -0.3448  1.5529  0.1873  0.0564
#>  0.7619  0.5867  0.6509 -0.3670  3.2437  1.7158  0.4049  1.0842 -0.7893  0.4271
#> -1.6250 -0.8630  1.0625  0.7039  1.9664  0.5898  0.3155  1.3787  0.2735  0.5270
#>  0.6682  1.2995  1.3111  0.6246 -1.1028 -1.0773 -0.5955 -0.5051  1.0237  0.7840
#>  1.8229  0.6374 -1.6985  1.4861 -0.1905  0.9070 -0.1736  0.8627  1.1633 -0.1226
#>  1.6086 -1.1342 -0.4723 -0.1605  0.0381  1.0691  2.7882  0.2665  1.3261  0.5227
#> -2.1944  1.3406 -0.7594  0.0516  1.7928  0.0738 -2.1911  0.1542  0.2340  0.9822
#> -1.0044  0.1165 -1.3286  1.0745  1.4210  1.4800  0.7109 -0.3706 -1.0486 -0.2159
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]