Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5462 -0.1474  1.0191  1.9058 -2.2198  0.7716 -0.7660 -2.2232 -0.2531  0.2946
#> -0.5487  0.3199  0.7180 -1.0705  0.1779 -1.5470  0.6737 -0.4268 -1.7108 -2.0936
#> -1.2538  0.4015  0.2089  0.5627  0.5475  0.5943 -0.3765  1.3445  0.2076  0.7974
#> -0.9493 -0.2491  0.1580 -0.2186  0.8374 -0.3438  0.5799 -0.7328 -0.5587  0.3117
#> -0.6978 -0.2172 -0.2908  1.5270 -0.0019 -0.3657 -2.1168 -0.3145  0.3142  0.1234
#>  0.4434 -1.1772  0.1322  0.3761 -0.8293  1.6255 -0.5512  0.6235 -1.0575 -0.3573
#> -0.1070  0.5780 -0.9806 -0.9268 -0.5391 -0.3265  0.0498 -0.1633  0.7551  1.5945
#>  0.3275  1.4104  1.3447 -0.0452 -1.3925 -0.1752 -0.6177 -1.4157  1.1859  0.7087
#> -0.4543  0.0512 -0.2524 -0.7729 -0.9755  0.6979  2.9726 -1.0729 -0.5120 -1.9146
#>  1.7841 -1.5512  0.6871  2.1790  0.4311  1.4066 -1.2454  2.0033 -1.3174  1.0526
#> -1.4552  2.3380 -0.0189  0.1812  0.6646 -0.2191  0.7724 -0.2336 -0.1776  1.3905
#>  0.7453 -0.0901 -0.7760  0.6536 -0.0921  1.8209  0.2547 -0.8120 -0.5388 -0.1112
#> -1.3588 -0.7383 -0.5764 -0.4379 -0.5067  0.1921  1.2742 -0.5078 -1.2427  1.5010
#>  0.9217 -0.5446 -0.7902 -1.0661  1.1863  0.5774  0.5371  1.8819  0.1622  0.2228
#> -1.3536 -2.0790 -0.3802  1.9156  0.9869 -0.2240 -0.1305 -0.1122  0.8489 -0.7095
#>  0.7678  0.6573  0.0092  0.3973  1.4259  0.1431  0.9054  0.4112 -2.4090 -0.9481
#> -2.1458  0.4015  0.8598 -0.3930  0.4345 -0.7660 -1.8893 -0.3247 -0.3391 -0.1502
#>  1.0932  0.5474 -0.6388  0.3202 -0.2544 -0.4802 -0.4026  0.6881  1.0089 -0.4208
#>  2.3109  0.8500 -0.6587  0.0769 -1.2506  1.2925  0.9910  0.2697 -1.1843 -0.1240
#> -0.0437 -1.4625 -0.3129 -0.5836  0.6115 -0.7612  0.1439  0.3647  1.2517 -0.6636
#> -0.4891 -0.0900 -1.0378 -0.8049  0.3075  0.4765 -0.0419 -0.3041 -0.6271  1.7379
#>  0.0141  1.4054 -0.2240  1.6352 -1.0852 -0.9448  1.8352 -0.9028  0.0789 -0.5768
#>  1.1959 -0.3144 -0.6080 -1.2453 -2.2452  0.9929 -0.0799  1.8169 -0.3511 -0.2491
#> -0.1895  1.1755 -0.4524 -0.5775 -0.7418  0.8686 -0.2683 -0.1063  0.2987 -0.6075
#> -0.4064  1.3328 -0.0582 -0.9169 -0.3527  1.4246  1.3455  1.7476 -0.7647  0.3296
#>  0.8131  0.5535  0.3459 -0.2588 -1.2908  0.0123 -0.8841  0.0921  0.9564  0.2547
#>  0.2833  0.6789  0.1775 -1.0299 -2.3634  0.1534  1.4728  0.4572 -0.2109 -0.3655
#>  0.2466 -0.1829 -1.0209 -0.1739 -0.5293  1.5040 -0.4364  0.1379  0.6213 -1.6003
#> -0.0332  0.7662  0.7768  0.0977 -0.7650  0.7889 -1.5239 -0.5731 -0.0781 -0.4291
#> -0.2117 -1.0865 -0.2202  0.4596  1.2448  1.1228  0.4955 -0.4095 -0.0741 -1.1887
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]