Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.1275  0.3063 -0.7041 -1.0469 -0.5376 -1.9947 -0.3430  0.1294  1.8186 -0.2541
#>  1.1276 -0.3920  0.4323  1.2878 -0.1073 -1.2117 -0.4775 -0.3898 -0.9685 -2.3865
#>  0.2960  0.9865  0.6621  0.0236 -0.7198 -0.8400  0.0863 -0.0933 -0.4971  0.3979
#>  1.8943 -1.4073 -0.3489  1.0887  0.3875  0.0723  0.6581  0.8854 -0.3364 -1.5274
#> -1.0166 -2.0017  1.2379  0.7418 -0.6280  0.8705 -0.4673 -0.1142 -1.0402  1.2485
#> -0.3961  0.2351  1.0809  0.7325 -0.0370  1.6884  1.0417  0.6857  1.7325 -1.6923
#>  0.2604  0.7646  1.0059 -1.0919 -1.1708  0.0467  0.0153  0.9204 -0.1739 -1.0016
#>  1.4067 -0.4723  1.0847 -0.2614 -0.8913 -0.3013  0.6162 -0.3795 -1.2601  0.6139
#>  0.4219 -0.0092  0.0179  1.1846  1.4082  0.8278 -1.4177  1.8998  0.4459  0.8972
#>  2.3840 -2.4220 -0.5996 -1.1804 -2.1790 -1.0688 -1.0960  0.0677  0.3926 -0.2895
#> -2.9539  0.1605 -1.5505  0.2657  0.5159  1.1819 -0.4038  2.2877  0.9828  0.7038
#> -0.0020 -0.6638  1.4665  1.3061 -0.1949 -0.9823  1.0980  1.4462 -1.6640  0.3234
#>  0.8655  1.1088  0.8225 -1.0186 -1.6163  0.7020 -0.7149  0.6324  0.6055  0.0238
#> -1.3545 -0.2750  0.0518 -0.6437 -0.6002 -2.0072  0.7448 -1.2167  1.1311  0.3279
#> -0.5104  0.2089  0.5348 -0.7949 -0.0122  0.3264 -1.1593  0.3829 -0.1806  0.0947
#>  1.1361  0.9655  0.3468  1.0748 -0.0946  0.2292  0.5845  1.8133 -0.1947 -0.7100
#> -0.7736  0.7801 -0.5071  0.9000  1.4160 -0.5680  0.3631  1.5270  2.0487 -0.2341
#>  1.4503 -1.3753  0.3443  1.1495  1.0319  0.2093  1.0661  0.5184  1.5519 -0.4130
#>  0.0177 -0.0795  0.4657 -1.0716 -0.2526  1.0803  1.1215 -0.4391 -0.4305 -1.0803
#> -0.7653  1.1334  1.6582 -0.2689  1.8265 -1.1699  1.1297  0.3830 -0.0470 -0.0102
#> -0.0242  0.8914 -1.3607 -0.6590 -1.1595  0.0989 -0.2678 -1.0293  1.3765 -0.5566
#>  0.4353 -0.6512  1.6256 -0.4275  0.4644 -0.4733  0.4610  0.5153  1.8640  0.8452
#>  0.7783 -1.5543  0.8977 -0.5961 -1.6867  0.9008  0.2167 -0.6962  1.1049  0.5144
#> -0.6966 -1.4535 -0.3179 -0.6151  0.1405 -3.3487  0.9389  1.6438 -1.1487  0.2806
#> -0.2957  2.4745 -0.6700 -1.1501  2.2606 -1.4607 -0.4355  0.2514 -0.2877 -0.4459
#> -1.7590 -1.3911 -0.6936 -1.1067 -0.3128 -1.7181  1.4126 -1.3921  0.1207  0.5642
#>  0.0154 -0.4633  0.7872 -0.8773  0.8286 -0.4122  0.6063  0.9398  0.7418 -0.2565
#> -0.6185  1.0566  0.7371 -0.2356 -0.2286 -2.1189  1.5450 -1.1758 -0.8681  0.2190
#> -0.6521  0.8440 -0.5913  0.3318  2.4194 -0.4251  0.2724  0.3826 -1.2476 -0.8956
#> -0.7416 -0.6560  0.1821  0.4217  0.0105 -1.3037  0.6716  0.2798  0.2926  0.3799
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]