Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.2307  1.2859  1.0579  0.7600  0.4781  0.5334  0.5665  0.9196 -1.2891  1.0919
#> -1.7643 -0.0700 -0.2692 -0.2086 -1.6699 -1.2782 -0.3846  0.5454 -0.2541  0.8373
#>  2.4716 -1.7960 -0.6008  0.8249  0.5447 -1.0360  0.3781  1.2664  0.6251  0.6119
#>  0.2334  1.1536  2.9661 -0.2057  0.9484  0.0891 -1.1794  0.7041  0.3935  0.8051
#>  1.6212 -0.6503 -0.7839 -0.0247  0.3107  0.2736  0.8982  0.4430 -1.0071  1.4593
#> -0.6129  0.6464  0.4643 -2.5563  0.0532 -0.3910 -0.0980 -0.2617 -0.1686 -0.3371
#>  0.8549  0.6711  0.2786 -1.0611  0.5025  0.2380  0.2908 -0.3329  0.6881  0.4006
#> -0.0554 -1.0021  1.6837 -1.2246 -1.2833  0.9925  0.1913 -0.5924  1.2289  2.0825
#>  0.1582  0.4925 -1.0289  0.5077  0.1213 -1.2803 -0.0991 -0.7403  1.1714  0.0048
#>  1.0247  0.0425 -0.9515 -0.3830 -0.5832  0.2435 -0.6071  1.1579  0.0703  0.2739
#>  0.5336  0.7819 -0.2729 -1.5300 -1.9983  0.2821 -2.0007 -0.1295  0.3992 -0.1334
#>  0.0618 -0.1093 -1.5318 -0.4775  0.2009  0.8613  0.7740  1.3531  0.2548 -0.3199
#> -1.0437  1.1811 -1.3534  0.3725 -0.5384 -0.1457  0.7524  0.5649  0.0681 -0.9768
#>  0.3969 -0.7355  0.2198 -0.0718  0.2532 -0.2313  0.5696  1.3239  0.1041  2.5084
#>  1.6193  0.8977  1.2774  1.7864  0.6661 -0.2813  0.8163  0.2477  0.8865  0.4888
#>  0.3603 -0.3288 -1.6945  0.3617 -0.4879  0.2704 -1.6085  1.3583  1.1335 -0.3626
#> -0.0715 -0.3924  0.8875 -0.8188 -0.5608 -0.9044  0.8627  1.3092  0.0475 -1.2094
#> -0.5560  0.7300 -0.2006 -1.4716  0.0941  0.1313 -0.6103 -1.8562 -1.7859  0.3925
#> -1.3484 -0.2752 -0.2196 -1.1537  0.8577  0.4589  1.2295  0.8149 -1.2965  0.4325
#>  1.6781  2.0261 -1.4911 -1.1611 -1.3584  0.5464  1.0410  0.6891  0.1100  0.6497
#> -0.5201 -1.6015  0.8259 -0.9284  1.3719 -1.5531 -1.1207 -1.5823  0.7930  0.8743
#>  0.1489 -0.1750 -2.0481  0.9415 -1.3810  1.0791 -0.1445  0.8962  0.4466 -1.0321
#> -0.3840  1.9073 -1.1689  0.1489  0.2657  0.0071  1.6338  0.0579  0.3145  0.3725
#>  0.5348  2.3341  1.4726 -0.0726 -0.2708 -0.7554  0.3763  1.0304 -0.5114  0.2923
#> -0.1191 -0.0863 -0.2979 -0.4081  1.6629 -0.4071 -0.8652 -0.6025  0.2538 -1.0398
#> -2.0176 -0.5905 -0.1521 -0.1227 -0.2017 -0.1481  1.3714 -0.8856 -0.2015 -2.1145
#>  0.2177 -1.1867  0.8388  0.1682  0.4162  0.2159 -0.8235 -0.3523  1.4970  0.2813
#> -1.1651  2.4877  0.9861 -0.5466 -0.5747 -0.2652  0.0713  0.0433 -0.5055  1.0268
#>  1.3100  1.2058 -1.5204 -0.6562  0.3158 -0.6620 -1.4649  2.7759 -0.1200  0.4570
#> -0.4666  0.2338 -0.2063 -2.0172  0.1113  1.1146  0.4092  0.3189  0.5392 -0.1792
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]