Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4038  0.5631 -1.3322 -1.4725 -0.2987 -1.3279 -0.0068 -0.4473  0.4675 -2.0615
#> -0.3722  0.3311 -0.5358  0.2060 -0.8980  1.3507 -0.1618 -1.3576 -0.7205  0.0887
#> -1.5871 -0.8113 -0.2681 -0.0953  0.6698  0.0976  0.2186 -0.4652 -0.9848  0.1239
#> -3.4095  0.2970  0.1051  0.0226  1.0896 -1.7646  0.1302 -1.2695 -0.0956 -0.8139
#>  0.5389 -1.2464 -0.8513 -0.1163 -2.8276  2.5888  2.3671 -0.2257  0.5544 -0.5270
#>  0.4879 -1.7984 -0.8986  0.7815 -0.0797 -0.1574  2.2233 -0.0920 -0.6382  0.1989
#> -1.0233  0.2563 -0.5482  0.0964 -0.6901 -0.7557 -0.2005 -0.4982  2.0095 -0.0576
#> -0.2914  0.1205 -0.9467 -0.7967  1.0803  2.5214 -1.6150 -0.1104  1.8408 -1.3944
#> -1.3774 -0.9629 -1.6052 -0.1898  0.5015  0.9995 -1.0616  3.1964 -0.7654  1.3823
#>  0.1262  0.1547 -1.0392 -0.3864 -0.6603 -0.7079  0.5606  1.1286  0.2106  0.2281
#>  1.0598 -0.0553 -0.6962 -0.0770  1.5530  0.7709  0.0650 -0.5595 -0.3673  0.0832
#> -1.4483 -0.4129 -0.3834 -0.2893 -1.0601  0.1243  0.0359 -0.4480  0.8565  0.7998
#> -0.1671 -1.6616  1.2635  0.4317 -1.5054 -1.6127  0.0669 -0.2362 -0.2962  0.3243
#> -0.7514  0.2884  1.4502  1.9782 -0.3495 -0.3182 -0.3550  0.3442 -2.7427 -0.9828
#> -1.3113 -0.3903  0.1947 -0.5100 -0.6258 -0.2810  0.5993 -1.6662 -0.6145 -0.9944
#> -0.6788 -0.9476  0.9833  0.5945 -0.6872 -0.2136  0.6594 -1.5510  1.3903  1.4308
#>  1.0990 -0.4829  2.0838 -1.0466  0.4182 -0.2732  0.7047  0.4776  1.1751  0.1149
#>  1.1783  0.9150  0.1370  1.0169 -0.4574 -0.1445  1.0456 -1.8012  0.6087 -0.2994
#> -0.6047  1.7299 -0.3029 -0.0555 -0.0462  0.4298 -0.0060 -0.3972 -0.6220 -0.6216
#> -0.5576  0.6569 -1.3800  0.7006 -0.7175  2.0292  1.1801  2.1572  1.4768 -0.7439
#>  0.6618 -1.3425 -1.7191 -1.6746  0.9412 -0.5413 -0.6602  0.3542 -0.8344  0.4985
#> -0.4266  1.4822 -0.2192  2.4225  0.6464  0.3032  0.0001 -1.0794 -0.6516  0.2855
#> -0.4022 -1.3541 -1.2894  0.5755 -0.1487 -1.7577  0.7604  0.7147 -0.7769 -0.6698
#>  0.4398 -0.2175  1.1821 -0.2122 -0.6593  1.1237 -0.9370  0.8765 -1.1364 -1.7737
#>  0.7577  0.7308 -1.0940  0.8666  0.3099 -0.8535 -2.8581 -0.0578  1.2216  0.0372
#> -0.2628  1.0775 -1.4290  0.2301  1.5918  0.2463 -0.9095  0.8746  1.0526 -1.1783
#>  0.2207 -0.3076 -0.5651 -0.8458 -0.4253 -1.2754 -0.0841 -0.6334 -0.1532  0.9982
#> -0.0496 -0.2474  0.7610 -0.7986  1.0741  0.7070 -0.9969 -1.6256  1.9313 -0.3843
#> -0.0186  2.3938  0.8802 -1.3078 -0.8824 -0.6540  0.2647 -0.3160  2.0011 -1.2480
#>  0.4290  0.6505  0.9577  0.4674 -1.1446 -0.0867  0.1799  1.6324  0.1194  0.3480
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]