Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7163 -0.2574 -0.3664 -0.0794 -0.5691 -0.3889  1.1567  0.3752 -0.8154 -0.7701
#>  1.2339  1.2287 -1.3956  1.1859 -0.7582  1.7797  0.5878  0.6514 -0.0175  1.0929
#> -1.0268  1.1678 -0.9383 -0.1889 -0.5787 -0.1834 -1.2095 -1.1771  0.4874 -1.4782
#> -0.4216 -0.4493 -0.2344 -0.7705 -0.4865  0.7805 -1.4842 -0.7400  0.0537 -1.9904
#> -2.5762  1.2057 -0.3849 -1.0665  2.1395  0.0547  0.9688  0.2734 -0.5595  0.4390
#>  1.4987 -0.1665 -0.9520 -1.3638 -0.3086  0.5215  0.7691 -1.8011 -0.4569 -1.3078
#>  0.9138  0.7102  1.6949 -0.3270 -0.8001 -1.7441  0.3279  1.7261  1.6571  1.2619
#>  0.4756 -0.3937 -1.2140  0.3909  1.5199  0.6925 -0.1894  0.4203  3.2852  1.3141
#> -1.1257  0.0138  0.7000  0.3410  1.3664  1.0632 -0.7405  1.2269  0.2293 -1.2095
#> -1.8261 -0.2670 -0.7018 -0.0025  0.1863  0.3847 -0.8287  0.4561  0.0080  0.4486
#> -0.8753  1.3273  0.5540 -0.0337  0.4846 -0.9497  1.4837 -0.2754 -0.5471 -0.0098
#>  0.5675 -1.2533 -0.7569  0.5056  0.2182  0.3648  0.0792  0.0390 -0.1462  0.1014
#>  0.7695 -0.8335 -1.2996 -0.5515 -1.4798 -1.3727 -0.9296 -0.0768  0.1298  1.0452
#> -0.1026  0.3185  0.5254  0.3225  0.4713 -0.5807 -0.0849 -0.9212 -0.5542  0.2276
#>  1.3839 -1.0252  0.0558  0.6664 -0.5049 -0.4999  1.2336 -0.0230 -1.2848 -2.0781
#>  0.8449  0.4859 -1.4364 -0.3467 -1.1000 -0.9515 -0.8670 -2.8017 -0.2062 -0.0380
#>  1.4540 -1.0558 -0.7189  1.8663 -1.0321 -0.5032  0.9220 -0.2738 -0.1409  2.7442
#> -0.3729  0.5913 -0.9235  0.5659  0.6845 -0.2004 -0.7550  0.1594  0.1758  0.7968
#>  0.9376 -0.4104 -0.7409  0.5491 -0.9157  0.4212  1.2412  0.0689  1.0736 -0.6188
#>  1.0595 -1.7104 -2.6967 -1.4960 -0.4590 -0.0613 -0.1630 -0.4666 -1.6736  0.5672
#>  0.3757 -0.9625  0.7375  0.2848 -0.0844  1.9045  0.2263 -0.9417  1.2050  0.4594
#>  0.2861  0.1046  0.0734 -0.6004 -1.1206 -0.8915  0.0803  2.5038  0.6452 -0.5042
#>  0.2455  0.7610 -0.4345  0.6730  0.5863 -1.5179  0.5097 -0.4898  0.0948  0.5317
#>  0.2151 -0.3823  1.2660  1.1276  0.2200 -0.3564  0.4145 -2.5612 -1.7572  0.9201
#> -1.7892 -0.4059  0.7812 -0.7002  0.3377  0.2786 -0.5561  1.5132 -0.0232  1.0044
#>  0.5256 -0.1598  0.4465  1.1669  0.6223  1.9105 -1.2857 -0.7394  0.5398 -2.8563
#> -0.8760  2.0643  0.9474  1.5465 -0.1622 -0.1514 -1.7966  1.6132 -0.0277  0.0357
#>  1.0472  0.1469 -0.7735  2.7905  0.9178  1.0259  0.1894 -0.3488 -1.0295 -1.1162
#> -0.4075 -1.5715 -0.5451 -0.2001  0.7345  0.3804  0.0319  1.0453 -0.3822 -1.2283
#> -0.5574 -0.2498 -2.1088  0.4342 -0.3339 -1.5715 -0.4542 -0.4800  0.6272  1.1964
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]