Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.2681 -1.6877  0.4001  1.8322  2.7276  0.9253 -1.5100 -0.0436  0.4871 -1.2473
#> -0.2177  0.1455 -0.1596 -0.1041  0.0699 -0.0697  0.2541  0.2108 -0.7315  0.2068
#>  0.2197 -0.3517 -1.3971 -1.0695  0.2687  1.3686  1.1431 -0.0529 -1.2919  1.0059
#> -1.1183  0.0349 -0.8760  0.3947 -2.4340 -0.8611 -0.9652  1.8648  1.0382 -0.1138
#> -1.1682  0.6891  1.3231 -0.3132 -0.4467 -1.1338  0.7550 -0.4933  1.7257  0.1049
#> -0.1679  1.0022 -0.0788  1.7964 -0.3713 -0.3198  0.4671  0.6655  0.4484 -0.5762
#>  0.5909 -0.1654 -1.4221  1.0742 -0.7105 -1.1246  1.9224  1.1205  0.6939  1.9193
#> -0.4293  1.8089 -0.3091  0.6146 -1.5901  0.9268  1.6312 -0.4256  2.0399 -0.2356
#> -1.2068 -0.0023  0.4553 -0.7865  0.6056  0.5545  1.2514  0.1350 -1.6453 -0.0796
#>  1.1850 -1.1980  2.3820  1.1588 -0.5651 -1.9754  1.7830 -0.8010 -0.5991 -0.7660
#> -0.2939 -0.0306  0.3579  0.5985 -0.3739  0.3804  0.5716  0.0983  1.2721 -0.0824
#>  0.2385 -2.2054 -1.5249  0.2282 -0.7283  0.4096  0.0995  0.8974 -0.6525 -2.0507
#> -0.7444  0.1960 -0.8214  1.1162 -1.2019  1.9448 -0.2133 -1.0594  0.0308  0.9570
#>  1.0006  0.3769  3.0977 -0.1608 -1.3036 -0.3848  0.9785  0.5545 -0.6769 -0.0176
#>  1.8237 -0.5406 -0.2306 -0.8450 -0.2958 -0.1183  0.6085 -0.2447 -0.0636  0.1647
#> -1.9192 -0.4403  0.3637 -0.2093  0.3992  0.3046 -0.4847 -0.6667 -0.2259  0.9579
#>  1.2663  1.2859  1.4227 -1.8654 -1.8276  0.4702  2.0570 -0.1471 -0.0951 -0.4822
#> -1.5853  0.1079 -0.8764 -1.6293 -0.5761  0.3906 -2.0962 -2.2508 -0.7402  0.1396
#> -0.4524 -0.4765  0.9794 -0.3791  0.1518  1.5742  0.5782  0.8354 -0.0869 -0.3853
#>  1.0295  1.0543  1.1681 -0.2113 -0.5574  0.2163  0.1427  0.3118  0.0716 -1.5232
#>  0.5095 -0.0076 -0.1485  0.1109  1.9781  0.0894  0.5973 -0.0933  1.5227 -0.3957
#> -0.7064 -0.9569 -1.1159  0.3458 -2.4043  0.2543  0.2251  0.4447 -0.9089 -0.5382
#> -1.4924 -0.8279  1.0839  0.5521 -0.3475 -1.0699  0.7247  1.3005 -1.0568 -2.6514
#>  0.4508 -1.2088 -0.4601 -0.3641  0.8697 -0.6423 -1.4814 -2.6329 -0.6259 -0.3479
#>  1.7569  0.0955 -0.9228  1.8474  0.3010  0.0132  0.8201 -0.8113  1.7361  0.3409
#>  0.5472  0.1035  0.4797 -1.5715  0.1510 -1.5698  0.5341  0.5880  2.2121  0.5897
#>  0.0087  0.9883 -1.5399 -0.6814  0.4588 -1.8289  0.6209 -1.6785  0.1974 -0.3628
#>  0.9144 -0.1291  0.8574  0.4355  0.0771 -1.8433  1.5459 -0.7589 -0.5870 -0.0308
#> -1.4774  0.9105 -0.3756  0.6939 -0.2993  0.9559  0.6810 -0.7040 -0.1887  0.3290
#> -0.6218 -0.2744  1.2124  0.8674 -0.6240 -0.2294  1.1510 -0.6221  0.4259  0.4938
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]