Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3291  0.8065  1.8887 -1.0687  0.1986  0.4149  0.8257  0.4596 -0.2810 -2.2655
#>  1.0902 -0.4532  0.5586  1.0747 -0.4004  0.7407 -1.4606  0.7790  2.0436 -0.9533
#> -1.0900 -0.4155 -0.2522  1.3543  0.4041  1.1801  1.2171  1.1912  0.9473  0.3341
#>  0.3471 -0.0971 -0.4756  0.8607  0.0591 -0.8772  0.1495 -0.3762 -1.3992 -1.6683
#>  1.9408 -0.5801 -0.4130 -1.2888  0.0474 -1.0571 -0.4720 -2.9304  0.2213 -1.9048
#>  1.8867 -1.8717 -0.1663  0.9996  0.9137  0.0399 -0.1007  1.2631 -1.3380  0.5327
#> -0.7354 -0.6289  0.4868  0.0301  1.2448 -0.2614  0.4275  0.2690  1.0935 -1.7197
#> -0.9638 -0.5672  0.0303 -0.2737 -0.2681  1.6507  0.8757  0.1528 -0.2890 -0.5509
#>  0.7692 -0.3969 -0.3979 -0.5413  1.6364 -1.4607 -0.3464  0.5306  2.6334 -0.7126
#> -1.7111  1.0819 -1.3692  1.1541  0.1351  0.7714 -0.1147 -0.0150 -1.1812  0.4740
#> -0.8174 -1.1509  0.3533  0.9021 -0.8849  0.7117  2.2146  1.1736 -1.0211 -0.8175
#>  0.5183  0.0525  0.0621  0.5385 -1.4567 -0.5946  0.3366  0.4376  1.1577  0.1427
#>  0.3522  0.9791 -0.1293  0.0955 -0.8293  1.0214 -0.7161 -0.1636  1.0620  0.8486
#>  0.3288 -0.6529  0.6276 -1.4059 -0.2047 -1.0723  0.9945  1.7017  0.6487  0.0642
#> -0.2551  1.5977 -1.8931 -0.7963 -0.2026  0.9964  1.0939 -0.7114  1.0142 -0.4931
#>  0.0786  0.9370 -1.9869  0.5289  0.1155 -0.4138  0.3481  0.3965  0.1024 -0.4679
#>  0.0807  0.7748  1.3568 -0.6516 -1.0986  1.4699  0.1638  0.4122 -1.5987  0.1518
#>  0.6198 -0.1016  0.4719  0.2908  1.8370 -1.4102  0.7045  0.1057 -0.1936  0.7763
#> -0.0437  0.4902  0.5024 -0.3296  0.5142  0.3073  1.2952 -1.4257 -0.8233  0.2748
#> -1.4411  0.0673  0.1666 -1.7794  1.6162  1.0178 -0.6725  1.0003  1.5996  0.5510
#> -1.8462 -0.0454 -0.5881  0.5931  0.1385 -0.9423 -0.4050 -0.0867  1.7561  0.8485
#> -0.1997 -1.3298  1.6858  0.4307 -2.5868  0.0757  0.6028 -1.1359 -0.0944  0.1903
#>  0.1227 -0.2541  0.2869 -0.4968 -0.2903  1.0493 -1.5299  0.2886  0.7543 -0.6339
#>  0.1349  1.1574 -0.3932 -0.3040 -1.6522  0.1093 -1.5959 -0.5433  0.5776 -0.9189
#>  0.6512  1.1116  0.0285  0.1817  0.1272  1.4145  0.1983 -0.2302 -1.1538  0.0931
#> -0.7870 -0.0256 -1.3961 -0.7461  0.3951  0.3486 -0.4693 -0.8434  1.1226  0.1288
#> -0.5171 -0.5432  0.1176 -0.3395  0.7480 -1.3590  1.0179  0.5564  2.1598 -1.4123
#>  0.3252 -0.3021 -0.6014 -0.1955  1.6699 -2.1950  2.0333 -1.6495 -1.0053  0.3023
#> -0.8057 -1.3421  0.0744  1.4918 -1.3545 -0.3327  0.2170  0.3816  0.4572 -0.1639
#>  1.4730  0.8027 -1.7588  1.5906  0.4139 -0.2306 -2.3060  0.3171 -0.2066  1.8447
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]