Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6825  0.3180  1.2597  1.6375  0.0423  0.3353  0.0507 -1.3449 -0.7900  0.9674
#>  0.6979  0.2996 -0.5715 -0.4914  0.2409  0.1248  0.1353  0.3460 -0.0505 -0.7411
#>  0.2573  0.7331  1.6356  0.0376 -0.5504 -0.0543 -1.8212  0.2507  0.7370 -1.6820
#> -1.0612  0.1836 -0.7047 -1.5176 -0.0245  0.4496  0.5812  0.0291  0.6418 -2.4772
#> -0.7268  0.3854 -1.1200 -0.0930  0.8567 -0.7414 -0.4783  1.3097 -0.0293 -1.6184
#> -0.4682  1.3530 -1.4949  0.6504  0.0999  1.2947 -0.6150  0.1739  0.1461 -1.6115
#> -0.6970 -0.5402 -1.1324 -0.0914 -1.2203  0.5446 -0.5493 -0.8766  0.3153  0.7314
#>  0.4288 -0.9477 -2.2030 -1.0100  0.9807  0.4030 -0.6150  0.6662  1.3047 -0.1927
#>  0.2838 -2.1254  0.0767  1.4362 -0.4462  0.2312  0.2598  0.9918 -1.5304  0.3697
#> -0.2932 -1.1277 -0.2337 -1.0637 -3.0336  2.4126  0.0729 -0.0051  1.0223  0.6652
#> -0.5695  0.3369 -1.2781  1.0605 -1.2984  0.8438 -0.8763 -0.2895  2.5836 -1.0870
#> -0.6906 -1.0750  0.9184 -1.1046 -0.4633 -0.9704  0.0889  0.4935  1.3515 -0.7535
#> -0.1008  0.8874 -1.0339 -1.3741  0.9820 -1.0333 -0.4860  0.6880 -1.0060 -0.1639
#>  1.0080 -0.4517 -0.2024  1.3726  0.3297 -0.0062  0.2211 -0.6108  1.3697 -0.3795
#> -0.4721 -0.0374  0.3124 -1.6578  0.6177  1.1717 -1.2111 -0.4279  1.6324  0.2054
#> -0.0640  1.1697  0.1434 -0.9849  0.3973  0.7123 -0.8217 -0.8213 -0.1944  0.4386
#> -0.8789  0.2423  0.4304  0.9835 -1.1957 -0.4830  0.5886  1.2771 -1.4213 -0.5585
#> -0.0056  1.2600 -0.2128  1.3283  0.5306  0.0994  1.3217 -1.4539 -0.2387 -0.5582
#>  0.7263  0.2144 -1.6117 -0.9860 -0.7528 -0.1535 -0.6013  0.3153  0.4256 -0.2353
#> -1.9214  0.4256  0.2298 -0.6955  0.3989 -0.3037 -1.5403  2.2073  0.6756 -1.7793
#>  0.5988 -0.4895  0.3528  1.5725  0.3564 -0.7303  1.0963 -0.4476 -0.5235  0.3705
#>  0.6469 -1.7379 -0.7268 -0.0554 -0.0066  1.0992 -0.2917  0.2999 -0.6643  0.8445
#>  1.0423  0.7229  0.8389  0.5808 -2.0409 -2.0236  0.8359 -0.2266  0.0061 -0.0487
#> -1.9408  0.5648  0.8224  0.5702 -0.7531 -0.6627 -1.5890  0.4897 -1.5710  0.7588
#>  1.2324  1.0389 -0.3119 -0.8279  0.4590  0.2030  1.4294  1.1154 -0.5537  0.7099
#> -1.6896  0.3104  0.1621 -1.4236  1.9120 -1.2044 -0.8502  0.0938  0.0885  0.6686
#> -1.9048  0.6757 -0.3497 -1.0288 -0.3281 -1.7563 -1.8864  0.5100  0.1332 -2.1420
#> -1.7725 -0.9317 -0.2473  0.7837  0.8551 -0.0316 -0.0634 -0.4188  0.0030  1.1463
#> -1.2814 -0.9095  2.5385  0.7164 -2.2095 -0.8511 -0.4945  1.5554 -0.9470 -0.3668
#> -0.0533 -0.3040 -0.8142  0.7634 -0.8679 -0.1591  1.0214  0.2417  0.0071 -0.1269
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]