Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.2137 -0.1446  0.1975  0.6958 -1.3456  0.6487 -1.0124  0.9544 -0.1751  1.6893
#>  1.2389  0.4638 -0.4730  0.7647 -0.0077 -0.5904  0.2157 -1.9320  1.2782 -1.0961
#>  1.5445 -1.2862 -1.6275 -0.0482  1.2276 -0.7121 -0.3145  0.4098 -0.4840  1.3182
#> -0.1134 -0.1814 -0.6903  1.8863 -0.9886 -1.1268 -0.6289  0.8031  0.1049 -0.6660
#> -0.6515 -0.3656 -1.3139 -0.0406 -1.0997  1.6078 -0.8016 -1.6226 -1.2916  1.2207
#>  1.4371 -1.0074  0.7799  1.7885 -1.5134 -0.1949 -1.9626 -0.8456  0.0767  0.3100
#> -0.1116 -0.8613 -1.1881  0.6905  0.8907  1.7909 -0.2496 -0.2237 -0.1071 -0.2295
#> -0.3538  0.6769  0.9286  0.3517  0.0197 -1.1037 -0.7662 -1.8336  0.5800  0.6339
#>  1.7102 -0.3648  0.5618 -1.1129 -2.1434 -0.3082  0.9981 -1.7607  0.3139 -0.8833
#>  0.2910  0.2021 -0.5735 -0.0148  0.7936 -0.0199  0.9824 -0.7396  0.9678 -0.1289
#>  1.6858 -0.1334  1.9568  0.4881  0.7767  1.2028 -1.7727  0.2405  0.1309  1.1704
#> -0.0790 -0.5753  1.3778  0.4414  0.7058  0.4860  0.0766  1.8136 -0.6233 -0.8805
#> -1.3244 -1.0138 -0.9216 -0.2074  0.1246  0.7111  1.5215  0.1526 -1.9501  1.2478
#>  0.1624  0.4842 -1.0608  0.3082 -0.0176 -1.3626  1.3975  1.0024  0.2107 -0.4937
#>  0.4049 -1.1247 -1.0730  1.6107 -0.6369  0.5134  1.9234  1.4304 -1.8785  0.1920
#>  0.4831  0.3878 -0.1581 -0.5670 -1.0881 -0.6191 -1.1117 -0.5988 -1.1545 -1.5456
#> -1.2186 -0.2618 -0.0510  0.0919 -0.5044  1.3942  0.9689  1.5477 -0.0160 -1.5493
#> -0.2891 -1.8095 -0.5202 -0.4060 -0.4126 -0.3311  0.0451  0.5511 -1.3288 -0.7537
#>  2.1357 -0.2315 -0.0942  1.6336 -1.8325 -2.4889 -0.1025  0.2463  0.1146 -1.3576
#>  0.0264  0.7793  0.6827 -1.8522 -0.2672 -1.3475 -0.2389  0.3285 -0.6292  0.8838
#> -0.5043  0.0246 -1.3673 -1.1809  0.2971 -1.1838  0.9325  1.4328  1.3282 -0.8736
#> -1.9829 -1.7504  0.6825 -0.0991  1.0639  0.2600 -0.2146  0.5409 -2.3905 -0.8341
#>  0.9477  0.0452  0.8299 -0.7976  0.3886  0.3541  0.8142  0.1856 -0.0057 -0.5281
#>  0.4657 -0.4776  0.8802  0.7705 -1.2930  1.9134 -0.5085  1.6619 -0.2828  1.6260
#> -1.1963 -0.8514  0.5977  0.6340 -0.7763 -0.0784 -0.3579 -1.5835  1.3730  0.9951
#> -0.2788 -1.0827  1.1557 -0.0904  0.6715 -0.4040 -1.5611 -1.5329 -1.9255  0.1153
#>  0.0237 -0.4746 -1.2808  1.0516  1.4815 -0.3299  1.1645 -1.7819  1.1691 -0.8243
#> -0.3030 -0.5775 -1.2918 -0.2623 -0.1313  0.1904 -1.0475 -0.3545  0.2509 -0.3675
#>  0.2930 -0.7448 -0.9973 -0.4545  0.5889 -0.1471 -3.2752  2.1617  1.4459 -1.6214
#>  0.1727  0.8544  1.7315  0.3908 -0.0152 -1.9015 -0.3751  0.4259 -1.3969  0.0827
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]