Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.1319 -0.2165  0.9383 -0.9056 -1.7317  1.9828  2.2076 -0.1265 -1.5911 -0.6240
#>  1.8055 -0.9246 -0.2484 -0.0708  1.2700 -0.3930 -0.8741  0.6907  1.3816  0.5046
#> -1.4556  1.4328 -1.7428  0.3780 -0.1618 -0.7195  0.4121 -0.6811 -0.8666  1.9465
#> -1.1235 -1.2053 -1.6542  1.8283 -0.2863  0.5901  0.1669 -1.4091  0.7192  1.7952
#>  2.7520 -0.6080 -2.1087 -1.4566 -2.6753  0.4262 -0.5365 -1.1243  0.8225 -1.1719
#>  0.0829  0.2505  0.0454 -0.3320  0.2228  0.7845 -0.4034  1.0979  1.6915  0.1305
#>  0.7631  2.1201 -1.4157 -0.8096 -0.2567 -0.1322 -1.1888  0.6758  0.0770 -0.3546
#> -0.5007  0.3021  0.8180  0.8961 -1.2822  0.3518  0.5934  0.1902  1.7378  1.5425
#>  0.6772 -1.3251  0.1988  2.3691 -0.6133  1.1692 -2.2059  0.3061 -0.0612 -0.5480
#> -0.3497 -0.2952  0.0485  2.1668 -0.2712  0.5572  1.1774  0.7744 -0.7655 -0.7051
#>  0.2589 -0.5610  0.6087 -1.8602 -0.7866 -0.7672  0.6921  0.5665  0.7002  0.2023
#> -1.8852  0.8539  1.3951 -1.5309  0.5718 -0.2338 -1.8070  2.2021  1.3030 -0.7189
#>  1.1110  0.9517 -0.5128  1.0521 -1.6837 -0.7025 -0.2731  2.4268  0.2381 -0.1114
#>  0.1386 -0.4762 -0.2950 -0.2501  0.3894  0.4051  0.5691 -1.1726 -0.8017 -0.0382
#> -0.8966 -0.4106  0.2446  0.2618  0.6425  0.4272  1.6706  0.3225 -1.1573 -0.9887
#> -0.6087 -0.4793 -1.1179 -1.1504  1.2890 -0.3801  1.1703 -0.9611 -1.1504  0.9215
#> -1.4110 -0.9855 -0.2200 -2.3977 -0.3929 -1.5702  0.6493  0.0090 -0.6213  0.2585
#> -1.5877 -0.5594 -0.4860 -0.4400  0.3646 -0.8057 -2.1149  0.2140 -0.9730 -0.5832
#> -1.1670 -0.4142  0.7174  0.7153  1.3785  1.1721  1.3304  0.1123 -1.1360  1.1393
#>  0.3994 -1.9939  0.1417  1.2612 -0.0582 -0.3445  1.0502 -0.3398  0.3659  1.0267
#>  0.1854 -0.8598 -1.0986 -1.3366  0.2879  0.1108 -0.0445  1.0700  0.4564  0.1638
#>  1.0568  0.3891  2.2468 -0.4986 -0.1242 -0.3954  0.7204  0.2293  1.6061 -0.2562
#> -2.0941  0.0979 -0.4216 -1.4821  0.9413 -2.4838  0.0039  0.7409 -1.7292 -2.3747
#> -0.1357  1.2767 -0.0489 -2.0965 -2.0229 -0.2762 -0.3382 -0.1716 -0.8419  0.9622
#>  0.1352  1.2520  1.0136  0.0241 -0.5516 -2.1884 -0.7130  0.7290 -0.5776  0.3379
#>  0.1764  0.3105 -0.7725 -0.7441  1.5756  1.4887  1.9199 -0.1417  1.3175 -0.7053
#> -0.7173  0.5814 -0.5449 -2.2876  0.4171  0.8756  2.2752 -0.6737  0.0652  1.2122
#>  0.0813 -1.0070 -1.4728 -0.2411  2.3968 -0.6257 -0.5343  0.9528 -0.2958 -0.6037
#> -0.3605  0.9376  1.0289 -0.7233 -2.9996 -0.6571  0.2523 -0.4230 -0.1059  0.5594
#>  0.3199 -0.1675 -0.6570  0.1185  1.2353  0.3724 -0.5980 -0.8611  2.6311 -0.6934
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]