Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7604 -1.3384  0.5968 -0.8441  0.6941 -0.2367 -0.4403 -0.7136 -0.6428  1.0400
#>  1.6158 -1.1853 -0.1723 -0.6828 -0.2970  0.3599  1.8331  0.1261 -0.2129 -0.4702
#> -0.5488  2.0884 -0.1944  1.4500 -0.7449  0.4170 -0.7980  1.0867 -1.3756  0.2513
#> -0.2946  0.2969 -0.5371  1.5497  0.6930 -0.3371 -0.7544 -0.4462  0.1245  0.7292
#> -0.5483  0.1887 -0.3653  0.5010 -1.0961 -0.8139 -0.5123 -0.2956 -1.5530 -0.3086
#> -0.7572 -0.3610  1.2853 -0.9795 -0.3040 -0.0971  0.1098  0.0748 -0.5647 -0.0159
#> -0.1368  2.0889  0.2144 -0.3745 -0.2558 -2.2538  0.7365  0.3291 -1.2515 -1.0517
#>  0.3722 -0.2490 -0.5137  3.5775 -0.2602  0.9455  0.6557  0.2451 -0.4916  1.2357
#> -1.5909 -1.1823 -1.5699 -0.0573 -0.3801 -0.0530  0.1826  0.1080  0.1540  0.9316
#>  0.8353 -0.0604 -1.4769 -1.5368 -0.6412 -1.1099 -0.5648  0.7094  1.1956 -0.5666
#>  0.0964  2.2644  0.7249  0.5180  0.2651 -0.1682 -1.0010  0.6408 -0.8418  0.0643
#> -0.5915 -0.1463  0.1075 -0.7690 -0.0326  0.0134  1.3333 -0.5121  0.4165  0.4347
#> -0.1506  0.5522  0.9301  0.5143  0.3545  1.0073 -0.1943 -0.2245 -0.6051 -2.2734
#> -0.9335  1.0242  0.9820  1.0370  1.1412  0.3126  0.1284  0.3077  0.6882 -0.6732
#> -0.5817  0.7075  0.3167 -0.1212  0.9119  1.2070  0.3339 -1.8339 -2.4164 -2.5135
#> -0.2425  1.3654 -1.2037 -0.8690 -0.4007 -0.2377  1.2205  0.8427  0.3457  0.2517
#>  1.3997 -0.8631 -1.0562  0.4387  0.9418  0.6133  0.3543 -0.6440  1.5710 -0.0262
#> -0.1224 -0.1820 -1.8651 -0.5827  1.4850 -1.8822 -1.1905  1.8663  0.2817 -0.1369
#>  0.6546 -0.2579 -0.0889  0.2368 -0.9404 -0.7320 -2.0205  1.6768  0.1366  0.4711
#>  0.1728 -0.0908 -1.8245  2.2531 -0.5185  0.7211  1.5543 -1.4493 -0.0387 -1.2313
#>  0.4884  0.8999 -0.7010  1.0492 -0.3466  0.5705 -0.1790 -0.0908  1.0789 -0.9119
#> -0.1547 -1.6604  2.4646 -0.0241 -1.1880  0.3334  0.1123 -0.0753  0.6056 -1.4611
#> -1.7123 -0.6945  1.3828 -0.2181 -0.5520 -0.1281 -0.4175 -1.6470  0.9815  0.6745
#> -0.3447  0.8280  1.0340  0.3263  2.0667  2.1226  0.8141  0.5749 -1.8717 -1.6668
#> -0.0180 -0.0956  1.4130 -1.5961  0.5975  0.3238  0.4279  0.6180 -2.1928  1.7452
#>  2.0999 -0.7228 -0.6220 -1.7200  0.2631  0.0849 -0.1015 -0.3206 -0.5426 -0.0015
#> -1.8009  0.8190 -1.1148  1.0009  0.7132 -1.1392  0.4797  0.5238  0.8219  0.2411
#> -0.5857 -0.3646  1.1431 -0.5920  1.0820 -0.2874  0.8084 -1.6068 -0.4301 -0.1972
#> -0.3120 -0.1781  0.4775  1.3346  1.6289 -0.5255 -1.1324 -1.9007 -1.0767  0.1359
#>  0.3776 -0.4003 -0.5825 -1.0695  0.0904  1.1155  0.5878  0.8342  0.3612  0.8196
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]