Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3772  0.4820  0.0596 -0.4040  0.1634  1.1406  0.1569  2.0604 -0.1580 -2.1436
#>  2.0176 -1.4039 -0.1326 -0.3289  1.5679  0.6180 -0.1605 -0.3665  0.8139 -0.7064
#>  0.4974 -1.5376 -0.5789 -1.3814 -0.4729 -0.0428 -1.2469  0.6335  0.9543  1.6963
#>  0.9252  0.1338 -0.5408  0.7408  0.1266  0.5015  1.0303  0.8416  0.1190  1.0851
#> -1.0998  0.1837  0.8470  2.3097 -0.3514  0.4342 -1.0801 -0.3815 -0.0381 -0.3357
#> -0.9127  0.3572  0.3623  0.2299 -0.3717 -1.9709 -0.0849  3.4281 -0.3435  0.1541
#> -0.2276  0.3914 -1.2742 -1.0630 -1.4170  0.5706 -1.9839  1.5774 -3.3991 -1.5238
#> -0.9315 -0.1688  0.2499  0.1361 -0.2981 -0.0257 -0.4480 -0.8543  1.6338 -0.2310
#>  1.9535  2.0792 -1.5728 -0.1041 -0.8992  1.5143 -0.5484 -0.4138 -0.4775 -1.5302
#>  1.0891 -1.2541 -0.2271 -2.6425  0.5483  2.3251  0.5672  0.0515  1.4574 -0.9424
#>  0.2381 -0.5641  0.6301 -0.8787  1.8327 -0.4217 -0.4961 -1.6742  0.0951 -0.5833
#> -1.6857 -0.8705  1.2899  0.6155  0.0414 -0.6058 -0.3775 -0.1393  0.0956  1.0781
#>  0.1112  1.4255  0.0215 -1.1600  0.7150  1.4863 -1.4275  0.1532 -0.2118  0.3521
#>  1.0706  0.4394  0.6087  0.7757 -1.4261  1.6789  0.4595 -0.8625  1.4629 -1.5495
#> -1.8248  0.3129 -0.0087 -0.1493 -0.0492  0.6379 -0.9952 -1.3087  0.5589  0.2525
#>  1.3281 -0.6284  1.1929 -1.9560 -0.7629  0.9823 -1.4824  0.1338 -0.5669  0.5377
#> -1.6444 -0.2557 -0.3961 -0.1668  0.6315  0.0069  0.6967 -0.0095  0.8883 -0.4226
#> -0.5771  0.9074 -2.0943 -0.3244  1.4790  1.2970  2.0446 -0.9238 -0.2720  0.5680
#> -0.0083 -2.1618 -0.1651 -0.5991  0.5519 -1.1776 -2.0534 -0.2393  0.5014  0.3271
#> -0.2253 -0.3775  1.3816 -0.5343  0.2078 -0.7782  0.5734  0.0580 -1.3402  0.2376
#>  0.5091  0.0625 -0.5501  0.2728 -0.0103  0.0346 -0.0842 -0.0280 -0.0857  0.4275
#> -0.1602 -0.4500 -0.8035  1.0451 -0.8969  0.2550  0.4575  1.0767 -1.8494  0.2048
#>  1.2322  1.2397 -0.9761  2.6934  1.2507  0.3569 -0.0711  0.4846 -0.4100 -0.1797
#> -0.3813  0.0971 -0.4988 -1.4772 -1.6285 -0.4666 -1.5444 -1.1044  0.0539 -0.5193
#>  0.4192  0.5175  0.5469  0.0035  1.6702 -0.8525  0.9962  1.0279 -0.4278  0.5906
#>  0.5859  0.6848 -0.1054 -0.2183 -1.0559  0.0338  0.7103  0.5844 -0.8391 -1.7863
#> -2.1614 -0.1844 -0.0488  0.2312 -1.0640  0.3267  0.4967 -0.4221  0.1786 -1.3755
#>  1.3726  0.5809 -1.4112 -0.8268 -0.6477  0.6856  0.3356  1.7131 -0.6398 -0.3748
#>  0.9313  0.1368  0.2761  1.4697  0.7343 -0.2247 -0.7358  1.8694 -1.8203 -2.3568
#>  0.5826 -0.1974 -1.9216  2.6422 -0.2736 -1.0196  0.9115  0.5312 -0.9466 -0.4292
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]