Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.1717 -1.0497  1.3232 -0.4998 -1.3977 -0.5338 -1.3323  0.1093  0.6636 -1.3267
#>  0.1949  1.7504 -0.6836  0.4327 -0.3829  0.9903  0.1261  0.5433 -1.7129 -0.1325
#> -0.9766  0.9855  0.9757  0.2295  2.4893 -0.6145  0.2085 -2.4949  0.5555 -1.0854
#>  0.1072  1.0100  0.1653  0.2497 -0.9469  0.2571 -0.2345  2.3847 -0.4791  1.3533
#>  2.2872 -0.0184 -0.1669 -1.4527  0.2455  0.4590 -0.6092  0.4311 -0.8137  1.2796
#> -0.5792 -0.7945 -1.7176  0.5561  0.2123  1.1910  0.9863  1.5869 -0.4888 -0.1381
#>  0.3371 -0.5491  0.7145  0.4602  0.0812 -0.2615 -0.4077 -0.4709  0.2888 -0.8919
#>  1.9263 -1.1380 -0.4477  0.2701  1.0460  0.4449 -0.7951 -0.6553 -1.6904  0.4311
#> -1.3561  1.3684 -0.4581  1.8971 -1.5182 -0.1764 -0.7871  0.1957 -0.3974  0.4549
#> -1.9224 -1.0198 -0.0322  0.5702 -0.0317 -0.6027  0.0895 -2.1229  0.8932 -0.8163
#> -1.5548  0.4619  0.9945  0.0610 -0.5657  0.5750  1.3111 -0.1031  0.7252 -0.5630
#> -1.2054  1.1050  0.5799 -0.4686  0.1575 -1.7822  0.0486  1.0441  0.7422  0.4681
#> -1.6083  0.3080 -0.7874 -1.0190  1.0602  1.3813 -0.2940 -0.1328 -0.1012  0.6229
#> -0.0968  0.7817 -0.6718 -1.6808  0.6197 -1.0321 -0.4143  0.4317 -0.2938  3.5915
#> -1.1947  0.7160  0.0607 -0.2862  0.1939  0.3243 -0.2544  1.2863 -1.4165  0.1303
#>  0.2915 -0.3415 -1.4080 -0.9721 -0.5722  1.7799 -0.5683  0.2168  1.1725 -0.4963
#>  1.1248 -0.6932  0.8682  1.0644 -1.5353 -0.1555 -0.2202  1.7477 -0.7984  0.6939
#> -0.7923  0.6091 -1.4184 -0.6524  0.4639 -1.9617  0.4443  0.8673 -0.6176  0.1267
#> -0.6227  0.5282 -0.1013  0.5473 -1.4945  0.9479 -0.4183 -0.0139  1.8820 -0.6644
#> -1.5249 -0.6007  2.1932  0.5074 -1.0571 -0.9862  0.6163  0.1805  0.3204 -0.9611
#>  0.3471 -1.4716 -0.4932  2.2923  1.8556  1.1478  0.9220 -0.1883  0.9387 -0.3322
#> -0.3858 -0.4680  0.7458  0.3409 -2.2808 -1.5041  0.2257  1.3980  1.9655 -0.6792
#> -0.4413 -0.1593 -1.8434  0.7428  1.1905  0.2285  1.6364 -1.8463 -0.6524 -0.3421
#>  0.5267 -0.0737  0.6500  0.6927 -0.3844 -0.2958  0.6278  1.5086  1.3709 -0.1372
#> -0.1324  0.7330  0.3423  0.5403 -0.7377  0.6844  0.0283 -0.7502  0.3527 -0.2096
#> -0.2400 -1.5681  1.3443  0.4503 -0.6705  0.1446  0.9265 -0.2289  1.2483  0.3665
#>  0.8391  0.0765 -0.7137  0.7316  1.6132 -0.4128  1.2215  1.2827  0.8125  0.5813
#> -1.9637  1.6494  1.8747  0.6900 -1.7169  2.6620  1.0058  1.0716  0.9840 -1.0384
#> -0.6459  0.8172 -0.7135  0.0862  0.4481 -0.0701  0.6188 -1.6385  1.0799 -0.8768
#> -0.6469 -0.1342 -0.3948  0.6251  0.0913  0.4052 -0.3203  0.4402  0.0393 -0.6927
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]