Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3567 -1.1490  0.6171 -0.2489 -1.3862 -0.5333  0.2855  0.0705  1.5725 -0.3126
#>  0.7938  1.5907  1.5599 -1.7392 -1.3812  0.2083 -0.4244  0.7154  1.0251 -1.1963
#>  0.2462 -0.3854  0.0410  0.4156 -0.1740 -0.0956 -1.6529  0.1640 -0.0778  0.7500
#>  2.0717 -2.5453  1.6750 -0.3388  0.0065  0.3504  0.4921 -0.7078 -0.6075  0.8839
#>  0.7000  0.9065  0.5470 -2.1325  0.6657  0.8106  0.5543 -0.0382  0.9126 -1.0199
#>  0.1418  0.9570  0.5659 -0.6927 -0.5571  0.6816  0.8084  2.6004  0.2120  0.5733
#> -0.8347  1.0621  0.6564 -0.8916  0.1983 -1.6444  0.5420  1.7954  1.0206 -0.6947
#> -0.6745 -0.3494  1.5327  0.3906  0.1330  1.0632 -0.6368 -0.9246  0.8217  0.3061
#>  1.2174  0.0252  2.0105  1.3942 -0.9103  2.2809  1.5622 -1.0918 -0.5364 -0.5707
#>  0.3528  0.7933  0.6158 -0.5749  0.7829  0.6415 -0.0669 -0.6498  0.2800  1.5507
#> -0.3316 -0.4012 -0.4937  0.3658  0.9915  0.5011  1.4848  1.4955 -1.3341 -1.4712
#> -0.2315 -0.2608 -0.2910  1.9619  0.9112 -0.8014  0.8699 -0.6912 -0.8252  1.0264
#>  1.2714  0.2589  0.9370  0.0352  0.7477 -0.1217 -0.7257  0.4502  0.3551 -0.6599
#> -0.0097 -0.3571  0.4603 -2.0829  0.3680 -1.2252 -1.1994 -0.4617  1.3190  1.1245
#> -1.3970 -0.3135 -1.0242  0.2602 -0.4687 -1.4781 -0.8678  0.0776 -0.2425  0.6193
#> -0.0528  0.2994  1.0546 -0.2114  0.2280 -3.2909 -1.0810 -0.9727  0.1320  0.9745
#> -0.0213 -0.5044 -0.6211  0.2873 -0.1790 -1.8841  1.5359 -1.7722 -0.1168  1.2848
#> -0.1104 -0.3053  0.8600 -0.9845  1.0467 -1.0529  0.0153 -0.5821  0.0266 -0.3854
#> -1.3573 -0.5287 -0.5373 -0.2651  0.8372  2.0789  1.3221 -0.4625  0.0096 -0.1706
#>  1.0790  0.2942  1.0061 -0.7698  0.5679  0.7072 -2.1779  1.7668 -0.3169  2.0177
#> -0.5427 -0.6010  0.6438 -1.0445  0.2352 -1.2214 -0.0246  0.8739 -0.2253  0.1933
#> -1.5655  1.0073 -1.6996 -1.6190 -0.8854  1.8236  0.1467 -1.4376 -0.4592  0.5717
#>  0.5717  0.6331  0.4702 -1.6731 -0.3726 -0.5783  1.2779  1.5634 -1.4925  0.0743
#>  1.2366 -1.6593  0.6351 -0.5066  0.1179  0.3021  0.7074 -0.4032 -0.0316 -1.9359
#>  0.2373  0.2716 -0.6660  0.9450  0.8670  1.9873  0.3506 -1.6672  1.0809  0.5628
#>  0.1952 -0.8059 -0.5473  1.6093  0.0470 -0.5323 -0.0068  1.1308  0.0737 -1.1469
#>  0.6713  2.0629 -0.5901  0.7282 -0.0660  0.8171 -0.1939  0.0496 -0.8109  0.2296
#>  1.3649  1.0447 -1.2759  0.4895 -0.7899 -1.5444  1.4413 -2.4530 -0.3852 -0.4900
#>  0.5552  1.0985  1.0806 -0.2250  0.5811  1.1549 -0.5200  1.5202 -0.1316  0.0413
#> -0.0995  1.4948 -0.6154  0.2022 -0.6254  1.1092  2.0351 -0.6669 -1.9234  1.6172
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]