Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0155 -1.7989 -0.0603 -1.4317  0.2272  0.5905 -0.1319 -0.8322 -1.7876 -0.8010
#> -0.0874 -0.6767 -0.9985  0.5712  0.7903  1.7494 -0.9217  0.4756  1.1873  3.1261
#>  0.5612 -0.0892  1.1337 -0.0230 -0.4148  1.0724  0.5335 -0.6938  0.9897  0.0543
#> -1.0382 -0.8295 -0.0505 -0.4645 -2.5706 -0.7290 -0.9875 -0.9263 -0.4053  0.1083
#>  0.0312  1.6880  3.2095 -1.4698 -1.2699  0.5419  2.5422  0.6075 -0.8460  0.1903
#> -0.7198 -1.5011  1.3739  0.8802 -0.9990  0.4353 -0.7417 -1.3328 -0.5492  0.1228
#> -0.2445  1.6875  0.7655 -0.1939  0.1718  0.0168  1.1637  2.1333 -0.6947  0.8572
#> -1.5189  0.6360 -0.7556 -1.0006 -0.2333 -2.2196  1.1889 -0.8756 -0.0123  0.4285
#>  0.5167 -0.6074 -0.4450 -0.2179 -1.6276 -0.9481 -1.5270  0.3139 -0.7688  0.6874
#> -1.2049 -2.3990  1.0953  1.2378 -1.1965  1.8235  0.3796 -0.2943 -0.1087  0.5373
#> -1.4819 -1.0165 -0.5317  0.9994  1.3718 -0.5905  0.0397  0.8840  0.2267 -1.8583
#> -0.1047  0.2311  0.4549  1.0016  0.6498  0.7167 -0.8181  1.2139 -0.6742 -0.4086
#> -1.1921  0.6509  0.4891  0.2410  2.0943  1.7670  1.2601 -0.7240 -0.6929 -0.5371
#>  4.2204  1.6429  0.3701  0.3364 -0.3858 -0.9276 -0.3128 -0.4708  0.4627  0.4944
#>  2.0710 -0.3017 -1.3939 -1.5670 -0.0580 -0.2076 -0.2366 -1.7603 -0.3536 -0.2034
#> -1.2991 -0.5030 -0.3720  1.6744  2.0046 -0.5851 -1.6615 -0.8598 -1.0777 -0.4169
#>  0.8199 -1.2905 -0.2045  0.7196 -0.6144  0.3865  0.6694 -2.6983 -0.3564 -1.2040
#> -0.4569 -0.3354 -0.3341  0.3280 -0.3125  0.3314 -0.9321 -0.9179  0.1550 -0.1254
#>  0.4168 -0.0374 -1.8886 -0.9822  1.0900 -0.1473  1.8183 -0.8767  0.7310 -0.2604
#> -0.8258  1.3105  0.5034 -0.7473  2.1129  0.9719 -0.0249  0.2360 -1.7778  0.0541
#>  1.1368  0.9364  0.1012 -1.7064  0.7577  0.0164 -0.2658 -1.5417 -0.2136  0.3326
#>  0.0250 -1.7673  0.2078 -2.2703  0.3724 -2.0406 -0.9491  0.3177  0.6568 -2.1491
#> -0.0592  0.6017 -0.4587 -0.4242  1.0701  1.4111  0.0564  0.1392  1.9664  0.2324
#> -0.8096 -1.7520 -1.2087  1.0120  0.3393 -0.1417  0.3999  0.5862  0.2571 -1.0014
#> -0.8074 -0.7852  0.7229 -1.4604 -0.8065  0.7497  1.3083  0.2022 -1.0840 -0.9329
#> -0.6559  0.3034 -1.0035  0.2030 -0.5791 -0.4307 -0.0537 -0.0720  0.1904  0.4045
#>  1.4389  0.7064  0.7188 -0.8128 -1.7096  0.2969 -1.4220 -0.9933  0.2857  1.4524
#>  1.8268  0.1909 -1.3157  1.4476 -1.0319 -0.8876 -0.2719 -0.8528 -0.6427  0.5637
#>  0.0258  1.5554  0.6487 -1.3262  1.6649 -1.3245  0.9468  0.8493 -0.7393 -1.3182
#>  1.0677  1.4089  0.0795 -0.0241 -1.2304 -1.0756  0.9863  0.0609 -0.0296 -1.4250
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]