Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.9517  0.0749 -0.6246 -1.2638  0.0810  1.6484 -1.4985  1.5258 -2.1677 -0.6146
#>  1.4637 -0.7063 -0.5492  0.1791  0.6654 -2.6499 -1.0062 -0.6685 -1.6290  1.2238
#>  0.0572  0.1307  1.1698  0.4470 -0.0561  0.0020  0.9378  0.5921 -1.1517 -1.5326
#> -0.6457 -0.7815 -0.9926 -0.6094  0.7598  0.3696 -0.2512 -0.1469 -1.3801  0.1928
#> -2.8209  1.2550  0.9725 -1.1090 -1.8371 -0.7916  0.4395 -0.9046  0.0059 -1.0677
#> -0.5053  1.8461  0.0803  0.9740  0.1858  0.5936 -1.4381  1.0617 -0.0476 -0.2093
#> -0.7601 -0.9452  1.4781  1.3200  0.9427 -0.6127  0.0141  0.8452  2.0758 -0.9112
#> -0.1934 -0.6077  0.3034  1.1453  1.1526  0.3370 -0.5558 -0.4893 -0.6510 -0.9002
#>  0.1053  1.2911  0.1741  1.0225 -0.2051 -0.6034 -0.9567  1.9703  2.3052  0.5105
#>  0.1485  0.0409 -0.8239  1.0013 -0.2386  1.8027 -0.6285  1.3377 -0.7720 -0.6330
#> -1.0308  2.0372  0.6962  0.9428 -0.2281 -0.8270  0.6919  0.8244 -1.0622 -0.7961
#>  0.1288  0.4540 -0.7072  1.0277 -0.9367 -0.8045 -0.9588  0.8565  0.5817  0.0696
#> -1.6215  0.6526  1.1327 -0.7663 -1.2269  0.8399  1.0871 -1.2574  1.1481 -1.6313
#> -1.1080 -0.3125  1.0007 -1.8720 -1.1258  1.5510  0.5619 -1.4680 -0.6379 -1.4011
#>  0.3460 -2.0747 -1.1015  1.2358 -0.5825  0.7956 -0.8342  1.0679 -1.6402  1.7037
#> -1.1210 -0.2263  0.1715 -1.3327 -0.4405 -0.2424 -1.5134 -1.9900  0.5861  0.8715
#>  0.2778  0.2561  0.8296 -0.1053 -1.0951 -0.6447 -1.8057 -1.9969  0.8917  2.2496
#> -0.6310  0.7337 -0.0912  0.3531  0.3509  1.0383 -1.0581 -0.2564 -0.5344  1.0478
#>  0.3696  0.5893 -2.3353  0.8471  0.5723  0.2245 -0.0172 -1.8288 -1.8348 -0.6798
#>  0.0776 -0.2100  1.2755  0.0365 -0.7135 -2.2070  0.0854 -0.6030 -1.7347  1.3554
#> -0.6617 -0.0521 -1.2615 -0.1820  0.5662  0.7899  1.1601  0.9815  0.2639  0.5113
#>  0.1772  1.8340  1.2245 -1.8906  1.3192 -0.6734 -0.8289  1.4091 -1.1017  0.0099
#>  1.2628  1.8341 -0.1735  0.7851  0.7009 -1.0649 -0.0816 -0.4363 -0.4447 -0.0686
#> -0.0199  0.6928  0.7422 -0.9625  0.3876 -1.8411  1.2992 -1.9473  0.9055 -0.3387
#> -0.0352  0.5659  0.0622  0.6107  0.4440 -0.1118  0.2790 -0.2658  1.0146  0.0745
#>  0.7063 -1.0673 -0.5012 -1.7610  0.4488  0.5699  0.0754  0.4262 -1.7170 -0.4944
#>  1.0107  0.1792  0.7598  0.1720 -0.2386  0.9807 -0.2224 -0.2032 -0.0956  0.1694
#>  0.1556  0.3370  1.4547  1.1190  0.7183  0.4487 -2.0859 -0.7464  0.3102 -0.8536
#> -0.7984 -1.0698 -0.3771 -1.2335 -0.5352  1.6548 -0.9104  0.0670  1.1794  0.9725
#> -0.5821 -0.7204  0.4713  0.8923  2.9071  0.0787 -0.4949  1.7271 -0.7165  0.0579
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]