Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.6257 -2.1841 -0.8143 -0.3420  0.8693 -1.0583 -1.0022 -0.5240 -0.0824  0.3968
#> -0.1157  2.0675  0.0295 -0.7145 -0.4251  0.4422 -0.3744  0.7137 -0.3644 -1.4799
#> -0.0826 -0.0631 -1.5655 -0.3429  1.0175  0.8776  1.7824  1.1134  1.1938 -1.3269
#> -1.3229 -0.2948  0.5305 -0.2207 -1.5386 -0.1216 -0.3165 -0.9798 -1.1753 -1.0949
#> -1.6096  0.7223  0.5643  0.1643  2.3977  0.6652 -0.3626  0.0017  0.4880  0.0121
#> -0.1224 -0.4205  1.5246 -0.7857  0.3648 -0.2607 -0.3220  1.5895 -2.2491 -1.0724
#>  1.1059 -1.4148  0.1131  1.1545  0.9565  1.2148 -0.4482  0.2741 -2.3697  0.3883
#> -1.1831  2.2296 -1.2191 -0.5522  0.9506 -0.3457 -1.2222  0.4827  1.6942 -0.4066
#>  2.1879 -0.6917 -0.0906  1.3179  0.9370 -1.6630 -0.2138 -1.4959  0.3245  0.9670
#> -1.6307 -0.6041  0.6934  1.5723 -0.0994  0.8997  0.6146  0.7544 -0.2415  0.5386
#>  0.2234 -0.8110  0.8504  1.1409  2.1412  1.9611 -0.3046  1.1728 -1.4987  1.2474
#>  0.0936 -0.5571 -0.5207 -2.5468 -0.2622 -2.1156 -0.8778  0.8236  2.6787  1.3306
#> -0.6199  0.3629 -0.1525 -1.0566  0.2327  1.6375 -1.5279  0.7462 -0.2799 -0.4335
#> -0.1454  0.6242  0.4486 -0.1576  0.2037 -1.5054 -0.6092 -0.2408  1.1838 -1.8075
#> -2.0963  2.9623 -0.6879 -0.7683 -0.6014 -0.6102  0.0820 -0.5280 -1.0139 -0.6854
#> -2.4378  1.6137 -0.0309 -0.7555 -0.4030 -1.6115 -1.1578  1.1925  0.3412  1.5726
#>  0.0209  1.1460 -0.3036  0.1867  1.0268  1.6465  0.9473  1.4650 -0.7237 -0.9621
#>  0.8881  2.4344 -1.4859 -0.4900  1.5397 -0.9946 -0.0670 -2.3709  0.8531 -0.0102
#> -0.2451  0.1829 -1.3861 -0.7687  0.3850  0.6703 -0.2896  1.0838  0.2722  0.2086
#> -0.8857  0.4983  2.7807 -0.8798 -0.5515 -1.6145 -2.0373  0.2889 -0.0020 -0.2156
#> -0.2836  1.3381  0.0454  0.2847 -0.1208 -0.2364 -0.8946  0.5642  0.5130 -0.2938
#>  1.9552 -0.5919 -1.1765 -1.2570 -0.2613  1.9554 -1.8556 -0.9881  0.8897 -1.1605
#> -0.9590 -0.9042  2.8327  1.2709 -0.5074  0.8657 -0.4431 -1.3671 -0.7684 -0.3032
#> -1.3857 -1.1421  0.6765  0.0952  0.4917 -0.6781 -0.1365  1.1230  0.4934 -0.6433
#>  1.2854  0.4860 -0.1722  1.3093 -0.4642  0.6708  0.0450  1.9961  1.5624 -2.4566
#>  0.3887 -0.7218  2.1599  1.4943 -2.8982  3.2217  0.2051 -0.2713  0.3898 -1.2398
#> -1.4641 -0.7057  0.6774 -0.9867  0.1576 -1.3278  0.4921  0.7816 -1.0775  1.1716
#>  0.4550 -2.3186 -0.7687  0.7548  0.4407 -0.7637  0.5342 -0.2277  0.4443  0.1885
#>  1.7620 -0.9930 -0.2971  1.1888  0.2289 -1.7960  0.5654  0.9183 -0.8434 -1.0580
#>  0.4315  0.2609  1.0874 -0.2320  0.8243  0.6609  0.7638  0.9323  0.3695 -0.2489
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]