Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.9830 -0.4516 -0.8371  1.8386 -0.7552  1.0966  1.2841 -0.4736  1.1786 -0.2090
#>  0.5679 -2.0209 -0.4452  0.1213  1.0943  1.2868 -1.1165 -0.0054 -0.6954 -0.9728
#> -1.6588  0.1609 -0.9648 -0.9381  0.0028  0.0413  1.4977  0.3451 -0.5334 -0.9522
#>  1.0570 -0.1996 -1.3718  1.7285 -0.9769 -1.3057  0.4010 -0.0807  0.1564 -0.2263
#> -1.8627  0.9264  1.3600 -0.7854  0.7755 -1.6210  1.6616  1.3039 -0.9519 -0.0570
#> -0.3951 -0.1861 -1.0426  0.3929 -1.0009 -3.0036 -0.2074  0.3847 -0.3635 -0.6626
#>  2.0978  0.7986 -1.3748  0.1896  1.5575  0.3346  0.3813  0.8413 -0.4059 -0.5172
#> -0.4950  0.0118 -0.7681 -1.7258 -0.3177  0.0437  0.7935 -0.1798 -0.3891 -0.6932
#> -2.1922 -1.3806  1.0329 -1.4901 -0.0428  0.7987  0.5740  0.2910  0.2361 -0.7316
#>  1.3221 -0.6817  0.4777 -0.1702  0.5520  0.4682 -0.8139 -0.3127  2.1832  0.4445
#>  0.4856  1.8010 -1.6733 -0.5690  0.9950  0.3566  1.2024 -0.0687 -0.6471 -0.6700
#> -0.7894  0.2053  0.1803 -0.6954  0.0965 -0.5121  1.1652 -0.7863 -0.0289  1.0009
#> -0.2416 -0.8078 -0.0718 -0.6070 -0.0949 -0.7097 -0.1598 -0.8597 -0.4628  0.0340
#> -0.3210  0.7108 -0.8879 -1.0780 -0.5602 -0.6812 -1.1358  0.3896  1.1996 -1.5362
#> -0.9905 -0.3997 -0.0623  0.0990 -0.2853  0.3322 -0.9255  0.9914  1.0149 -0.4496
#>  0.7449  0.4352 -0.3055 -1.2555 -0.5608  0.2019 -1.3586  0.8773  1.1150  1.4799
#>  0.1193  1.1573  0.0947 -0.5793  0.2053 -0.4251  0.6125 -0.1760 -1.1491 -1.5210
#>  0.7635 -0.6760 -0.2508  0.5266 -0.1558 -1.2479 -0.3223  0.4941  0.1065 -0.9871
#> -0.4230 -1.2564  0.5083  0.6902 -0.4412  1.4710  1.0106 -0.6235 -1.6592  0.8032
#> -0.1971 -0.0726 -0.5280  0.4344  1.0001  0.0948 -1.3472  0.6367  0.6567  0.5779
#>  0.3869 -0.7674  0.6720 -0.0555 -3.2464  0.6151  2.2674  0.5871 -0.2438 -0.7786
#> -1.0054  0.8195 -0.2730  0.3895 -1.2105  0.3831  0.5056  0.2023 -0.6137 -1.4623
#>  0.4215 -0.7156  0.7392  0.3426 -0.5966  0.4934  0.3034 -0.2720  0.0269  0.5457
#> -1.4821 -1.3611  1.3876  1.0129  0.5977  0.3684 -0.8450  1.4188  0.3309  1.7889
#> -0.9122 -0.1936 -0.1342 -0.7454 -1.5210 -0.2074  0.9427  0.6202 -1.8218  1.5490
#> -0.4002  0.7078 -0.3361 -0.6207  0.7292 -0.2684 -1.8972 -0.1722 -0.8050 -0.3557
#>  0.4044 -0.0861  0.9691 -1.4146 -0.6688 -0.5678  0.1778  0.7853 -2.8828 -0.3482
#>  0.4248 -0.9082  0.6367 -0.6805  0.9681  0.6823  0.5872 -1.4856 -0.3389  1.6628
#>  2.1268  1.9614 -0.5436 -1.4355 -0.0125 -0.3633  0.6437 -0.5254 -0.5228  0.9511
#>  0.2426  0.8206  2.7717  1.7410  0.8932 -0.7034 -2.3618 -0.1978  0.2491 -0.7592
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]