Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 1.7044e-01 -9.5092e-01 -9.3928e-01  2.2795e-01 -9.9876e-01 -2.9931e-01
#>  9.5155e-01 -1.1444e+00  3.7595e-01  4.5203e-01  1.2725e+00  1.7354e+00
#> -2.1409e-01 -1.0477e+00 -3.8247e-01  1.8427e+00  2.8707e-01 -1.1672e+00
#> -8.1009e-01  1.4337e+00 -6.1256e-01  7.4991e-01 -8.1924e-01 -6.2964e-01
#> -4.6302e-01 -5.1743e-01 -3.1498e-01 -1.0427e+00 -1.8775e+00 -9.6824e-01
#>  5.0992e-01 -9.2784e-01  7.8135e-02  1.9169e+00 -1.7788e+00  3.1428e-01
#> -1.7360e+00  4.8636e-01  5.3777e-01  1.5103e+00  3.4637e-01 -8.2009e-01
#> -7.0001e-02 -5.8115e-01  8.0791e-01 -1.0702e+00 -9.5559e-02 -1.0033e+00
#> -2.8545e-01  2.3960e-01 -1.2256e+00  1.5948e+00 -4.5864e-01  1.4690e-01
#>  4.9685e-01 -1.3377e+00  1.1675e+00  5.5999e-01 -1.2974e+00 -1.0683e+00
#> -1.0527e-02  9.0063e-01  8.7102e-01 -6.2805e-01  1.0061e+00 -7.4937e-01
#> -2.0096e-01 -9.1264e-01 -3.4991e-01  7.5011e-02 -3.2283e-02 -1.3589e+00
#> -5.9280e-01 -2.4144e-02  2.2618e-01 -1.2239e+00  6.7164e-01  3.1017e-02
#> -4.6319e-01 -1.5145e-01  1.9976e+00 -7.9280e-01  1.0701e+00 -9.4010e-01
#>  6.6315e-01  6.5543e-01  3.2296e-01  1.2830e-01 -1.1504e+00  3.5971e-01
#>  2.0615e-01  4.0687e-01  1.4553e+00  1.4160e-01 -2.0215e-01 -8.9081e-01
#>  1.2346e+00  6.6885e-01 -9.7882e-01  1.9250e+00 -1.0680e+00  7.4425e-01
#> -9.7994e-01 -1.1104e+00 -5.9304e-01  1.3504e-01 -1.3887e+00 -6.4090e-01
#> -9.5815e-01 -3.7000e-01 -1.1349e-01  7.0427e-01  2.3916e-01 -1.6929e+00
#> -4.4316e-01  1.7357e-01  1.7269e+00  9.0952e-01  8.9287e-01 -1.1889e-01
#>  3.8993e-01 -9.1667e-01  2.0052e-01  1.3418e-01 -1.2306e+00  1.1903e+00
#>  5.6826e-01  8.2293e-01  1.2202e+00 -2.6981e+00 -3.0124e-01  1.1565e+00
#> -2.9112e-01 -4.3002e-01 -5.2084e-01  1.0031e+00  3.4119e-01 -8.3125e-01
#>  1.1065e-01  9.2974e-01  1.1844e+00 -1.2689e+00  2.0280e-01 -3.7184e-02
#> -2.0265e-01 -5.6507e-01 -3.1052e-01 -4.5095e-01 -9.2261e-03  2.6367e-01
#> -7.5549e-01  6.9310e-01 -1.6076e-02  3.4832e-01  6.1503e-01 -1.0499e+00
#> -1.1105e+00 -7.1145e-02 -4.2537e-01  1.7546e-02  4.2721e-01 -3.9465e-01
#> -7.4046e-01 -8.4050e-01 -1.2368e+00  2.3737e-01  2.1531e+00 -2.9732e-01
#>  1.2589e+00 -2.5933e-01  8.4094e-01 -1.0372e+00 -2.7189e-01 -1.9452e+00
#> -7.9433e-01  1.4360e+00 -4.8194e-01 -9.2266e-02  5.3351e-01  1.3254e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]