Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.5654  1.5265  1.6618  0.9217  0.0313  0.1045  1.3879  0.0537  1.5736 -0.2828
#> -0.9829 -1.3091  0.8656 -0.3526 -1.2417 -1.4547  1.9209 -0.3158 -1.3853 -1.5252
#>  0.8796  0.3692 -1.0811 -0.0229  0.1137  0.5309  0.5783 -1.4009 -0.5420  1.8068
#> -0.7423 -0.2850 -2.1850 -0.8954  2.1577 -0.7895 -0.0636  0.6848 -1.4133  2.0980
#> -0.8848 -0.3938 -0.1738 -2.0020  0.6646 -1.1397 -0.9226 -2.8473  1.6356 -0.9912
#> -0.2937 -0.0450  0.3602  0.5960 -1.7292  0.0705 -1.1073 -1.7205  1.4264 -1.3680
#>  0.3558  0.3853 -0.4615  1.6681 -0.4546  1.2960 -0.6720 -1.0057  0.2856 -0.1224
#> -0.1632 -1.1554 -0.8162  0.0380  0.8522 -0.1019  0.3571 -0.6470  0.3311 -1.0433
#> -0.0530  0.0359 -0.0288  0.5259  0.0812 -0.4568  2.2484  0.8785  0.4805 -0.7991
#> -2.2125  1.1712 -1.0596  0.8889 -0.0289  0.1558 -0.5721  0.4147  1.9677 -1.8390
#>  0.1265 -0.7842 -2.0095 -0.8277  0.7397 -2.0773 -1.1800 -0.6965 -1.8001  0.9949
#> -0.6036 -0.4623 -0.8171  0.3121  1.2231  1.0402 -0.0910 -1.5794  0.0562  0.7524
#>  0.7968  0.8215  1.0840 -1.3329 -0.7973  1.6123  0.6510 -0.5003 -1.1403  0.3305
#> -0.4549 -0.3325  0.6067  0.6363 -1.5033 -1.2605  1.8653 -1.1909 -0.3534 -0.7181
#> -1.0555  0.6060  0.0597 -0.7600 -0.3858 -1.2667 -0.6125  1.2887  0.0163 -0.5141
#>  0.3873 -0.1149 -0.0318  1.1273  2.2900 -0.7925 -0.2834 -0.1044  0.4121 -1.5101
#> -1.4548  0.2806  0.2922 -1.5772 -0.8677 -0.2788 -0.5210  0.6234 -0.7591  0.1623
#> -1.0352  0.1978  0.4209 -1.9717 -0.4813 -0.3549  0.2783  0.4297  0.7175 -0.6849
#> -0.1174  0.4669  1.9506 -0.0699 -0.4928 -0.2535 -0.3293  0.0120  1.3425 -0.4156
#> -2.0803 -1.0077 -0.3650 -0.0832 -0.2840 -0.6689  1.1973  0.7263  0.1639  0.8537
#>  0.3970  0.7047 -1.8038 -0.3240  1.7932  0.2864 -0.6651  0.0859 -0.2146 -0.9129
#> -0.9317  0.8953 -0.8241  1.2157 -0.1108  1.8316  2.0772 -0.8084  1.3781  0.4818
#>  0.4221 -0.4139  0.1879 -1.6385  0.3762  0.5414 -0.9720 -1.0020 -0.8062  1.1198
#> -0.3380 -0.2411  2.0100 -0.0580  0.0104  0.7332  0.5386 -0.6052 -0.3404 -0.4737
#>  0.0984  0.0865 -0.2676  0.3433  1.0935 -1.4590  1.0392  0.1246  1.8121 -0.2297
#>  0.8348 -1.3930  0.6011 -0.2136  0.5794 -2.1771 -0.1333  0.8172 -0.4881 -0.1924
#>  0.7137  0.4571 -0.2676  0.8129 -1.0804 -0.3091  0.4556  0.4068 -1.1738  0.0406
#>  0.4237  2.3769  2.3232 -0.1927  0.2895 -0.8784 -1.0063  0.7612  0.5391  1.0589
#>  0.0593  0.4525  1.1014  1.5420 -1.3800  0.8414  0.9768 -0.6183  1.6634  0.5575
#>  0.2279  2.6016 -0.1679 -0.9894 -0.5754  1.4511  0.7567 -1.1404  1.2378  0.5379
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]