Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 2.2952 -0.3837  1.3767 -1.5990 -1.0475  2.8344 -0.1113 -0.0066  0.2099 -0.3617
#>  0.5027 -0.1758 -0.0022 -1.9035  0.4977 -0.5472 -2.6595 -0.3226  1.2970 -0.7790
#> -2.5618 -0.6857  0.2494  0.4017 -1.8368 -1.2583 -0.0826  0.5753  0.1660  2.6394
#> -0.0795  0.0140 -2.8541  0.1843 -1.0972 -0.4189 -0.5805 -0.1894 -0.5326  1.2457
#>  0.1453 -0.6366 -0.0157  0.0251  0.6915 -0.6057 -0.6363  1.5744  0.1358 -0.3227
#>  1.1120 -0.3793  1.3833  0.1244  1.5973  1.0932  1.6677 -0.7836  0.7991  1.1475
#> -0.1591  1.4260 -0.2941 -0.4809 -0.1231 -0.7783  2.2336  1.8315  0.1736  1.5063
#> -0.1067  0.6443  1.3769  0.4964  0.6681 -0.4915 -0.2228 -0.6512  0.6238  0.7856
#> -0.7672 -0.1033  2.2428  0.7950 -1.5289 -1.5534  1.2037 -0.4670 -0.4192 -0.9671
#>  0.5604 -1.2420 -0.6204  1.0613  0.9490 -0.0738  0.1189 -0.4285 -0.6592  0.2061
#>  0.0993  1.1986 -0.8846 -0.8600 -1.1066  0.1864  0.0913 -0.6035  0.9838 -1.2374
#>  0.1672 -1.3153 -1.0463 -0.8850  0.2511 -0.0005 -0.2262 -0.3278  1.5515  0.3614
#> -1.2407  0.2493 -0.3448  1.7452  0.2523 -0.9655  0.7365  1.0201 -0.1601  1.0618
#>  1.8826 -0.5334  0.3975 -0.2991 -0.4302 -0.6799 -1.0756 -1.6647 -0.0257 -1.6454
#> -1.3279 -2.1106 -1.3422 -0.6343  0.0066  1.6260 -1.4084  0.3139  0.5730 -0.7139
#> -0.4526  0.1846  0.0511 -0.3592  0.8895 -0.0355 -0.5547  0.5859  0.6562 -0.2834
#>  1.5520 -1.3947  0.8092 -0.6385 -1.5283 -1.7882  0.9365 -0.3873 -1.8544 -1.5194
#>  0.3874 -0.3749 -0.4772 -0.8212  0.5103  0.0107 -1.4117 -0.2252 -0.0645 -0.1396
#>  0.7134  0.0815  0.8306 -2.4042  0.8626 -1.1853  0.5073 -0.0855  0.5912 -1.0624
#> -0.6327  0.0797 -2.0984  0.5681 -1.3525  0.2025  0.2000  1.6393  1.2420 -0.0727
#> -0.3256  0.1830  0.4079  0.9230  1.6388 -0.7564  0.2651  1.1987 -0.1371 -0.6040
#> -1.0971  1.4660 -0.5078 -0.8729 -0.1708  0.1596  1.0236 -0.2147 -0.1398 -0.9834
#>  0.3180  0.2219  0.7513  1.0271  0.1446 -0.7225  0.5608 -0.1099  0.4308  1.0846
#> -1.7129 -0.7891  0.8928 -0.1253 -0.1686 -0.2542  1.2507  0.3190  2.0077 -1.0498
#> -0.0472 -0.4836  0.4339 -0.4953  0.1708  0.7166 -0.4620  0.3014  0.0188  0.8132
#> -0.0248 -0.5822 -0.9808 -1.2477  0.1047 -0.7316  0.8971  1.0040 -0.2092  0.0021
#>  0.3439  0.9053  0.2127  0.1655  1.1340  1.3762 -0.5587 -0.5150  0.9059  0.2286
#>  0.1772 -1.1609  0.0895 -0.4002 -1.6439  0.9495 -0.2725 -0.6565 -1.6373 -1.4850
#> -0.6716 -0.5676  0.7600 -0.4111  0.5674  0.4057 -0.8831  0.9220 -0.3083 -1.6833
#>  0.2941  0.4189  0.4462  0.4852  1.4647  0.9824 -0.5552  1.3490  0.7831  0.6034
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]