Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.7848  1.0489 -0.1646 -0.9574  1.0674  0.6078 -0.7038  0.9268  1.0313  0.4107
#> -0.1590 -1.7822 -0.9309 -0.0977  0.9794  0.8171  0.2283  1.1067  0.3774  0.9670
#> -0.7481 -0.1176 -0.6022 -0.4328  0.8188 -0.8562 -0.4939  1.5311 -1.6480 -0.1541
#> -0.6623 -0.1470  2.6742 -0.8140 -0.4477 -1.4332 -0.2505 -0.2235 -2.4067 -0.4746
#>  1.1763  0.1665 -0.5354 -0.9137 -0.1033  1.6951  0.9463 -0.8888  0.2684 -0.5287
#>  0.4457  0.2570 -0.2940 -0.7016  0.5242 -1.0892 -0.8159  0.1437  1.2293 -1.0611
#>  1.1641  0.8947 -0.3499 -0.4718 -0.9437  0.5666  0.8529  0.9235  0.2031  0.0084
#> -0.3993 -0.4771 -0.6299  0.4666  1.2784 -0.1018 -1.7841 -0.4978  0.2673 -0.5925
#> -1.3952  3.6578  2.5468 -1.2106 -0.5353 -1.1701  1.0826  0.6037  2.3021 -1.6943
#>  0.6051 -1.1230  0.1547 -1.3071 -0.5454 -0.6890  0.8115 -2.2366  0.7592  0.6975
#> -0.2384  0.5707  0.8685  1.1573  0.5514  1.4578  0.2916  0.7884 -0.9056 -0.3043
#> -0.4733 -1.5688 -1.9339  2.0829 -0.4181  0.5972 -1.5852 -0.1161 -0.4734  0.6860
#> -2.2879 -0.2426  0.9917  0.3609 -0.4603 -0.2529  0.0406  0.5292  0.7256 -0.9906
#>  0.7316  0.4907 -0.5874 -0.5811 -0.8053  2.2116  0.2746 -1.6731 -0.4572 -0.2597
#>  0.4794 -0.3832 -0.9418 -0.8965  0.3005  0.4087 -1.4834 -0.6697 -1.1588  0.9669
#> -0.2528 -2.4755  0.6237 -1.7898  1.0194  1.8569  0.0673 -1.3636 -0.8187 -2.5533
#> -0.3673 -1.3128 -0.3776  0.8390 -1.3352  0.6828 -0.4532  0.0599  1.3883  1.2655
#> -0.3390  0.0257 -1.0822  1.7037 -1.0526  0.2025 -0.1361  0.0357 -1.2840 -0.5186
#>  0.0899 -0.8277 -0.2797 -1.3310 -0.9411  0.4661  1.4461 -1.0522 -0.2905 -0.2953
#>  0.4875  0.4472  1.7980  0.5627 -1.7381  1.2507  0.3751 -0.6679 -0.7622  0.2851
#>  0.3157  0.2938  2.0423  0.4291  0.1610 -1.3645  0.7277 -0.0256 -0.0343  0.2608
#>  0.1451  2.8655 -0.2874  0.1165 -1.2834 -1.6955  0.8846 -1.3954 -0.1600 -1.2156
#>  1.1246 -1.6602 -0.6228  1.3135  0.0002 -0.9863  2.2884  0.7421  0.8154 -0.3985
#>  0.6200 -1.0895 -0.1674 -0.1454 -2.5294 -1.1591 -0.4761 -0.0713  0.0747 -1.6618
#> -0.6430 -1.5923  0.8223  0.4599 -1.3223 -0.0746 -1.2261 -0.7486 -0.0686  1.5782
#> -2.3490  0.7057  0.0786 -2.0034 -1.7090  0.3481  0.8433 -0.9607 -1.2346 -0.2980
#>  1.5123 -0.0396  1.0395  0.2608 -0.4660  0.0338  0.3120  0.4043 -0.0104  0.0019
#>  0.3791  2.0041  1.1338 -0.7237  0.1685 -0.8187 -0.0487  0.3700 -0.1746  0.1081
#> -1.2747 -0.9564 -0.8958  2.4006 -1.3575 -0.4682 -0.4343  0.1908  2.2201  1.0706
#>  0.4605  0.3618  0.6515 -1.0099  1.5199  0.2165  1.9051  1.7842  0.1839 -0.0630
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]