Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2998  0.7760 -0.1485 -0.4739 -0.5188  1.7405  0.5574 -1.7473  1.1024 -0.0301
#> -1.1211 -0.3811 -1.1643 -0.9918 -0.6212  0.5279  1.6509  0.2769  0.4949 -1.0684
#>  1.0416 -0.5419 -0.6515 -0.8026 -0.4372 -2.0763  0.5315  1.2380 -1.7542  0.6189
#> -1.2995  0.5049  1.5893  0.3239  1.1245  0.2614 -1.1793  0.4735 -2.4822  1.3780
#>  0.9926  0.0673  1.0191  0.3229  1.1746 -1.3698  0.2222  0.3803  0.9492  0.9046
#> -0.3105  0.1111  0.3746  1.2243 -0.5344  0.2598 -0.7333 -1.0398 -0.9636  1.1002
#>  2.1266  0.2394  1.5573  1.0384 -0.6846 -0.8979  0.2433 -0.1410  0.6526 -0.4806
#> -0.3660 -1.9039  0.0346  0.4083 -0.5106 -2.2182  0.8771 -0.3671  0.2209 -1.9531
#>  0.2446 -0.5525  1.0893  1.4829 -0.1342 -0.0889  0.5361 -0.0056 -2.1482 -0.2818
#>  1.2538 -1.5905 -0.9005  0.7628  2.3739  0.8766  0.7573  0.2128  0.6278 -0.0989
#>  0.6549  0.8689 -0.2725  0.2373  0.0073 -0.9058  1.1968 -1.3083  0.1884 -0.1782
#>  0.3461 -0.0193  1.2221  0.4808 -0.4763 -0.4180 -1.6940  0.5735  0.0782  1.3587
#> -0.0233  1.4883  1.0255  1.7779  0.3932 -0.4299 -0.0817 -0.3784  2.0036  1.1868
#>  0.7963  0.9312  0.3188  0.3315  0.8232 -0.5376  0.9579  0.9685  0.3240 -0.5583
#>  0.7201  1.3036  0.6051  0.3509  0.9825 -0.9359  1.5446 -0.6920  0.7700  1.2945
#>  0.0606 -2.2053 -0.5393  1.0913 -0.5521  0.0279 -1.6693  0.8486 -1.0148 -0.9768
#> -0.9987 -0.9618 -0.9998  1.4249  0.8542  0.2882 -0.5041 -0.1876  0.1729 -0.0515
#> -0.0394 -0.4204 -0.3140 -0.5614 -0.7032  0.2757  2.1194  0.5709  0.0527  0.1412
#> -0.7277  0.0461  0.1354 -0.5607 -1.3360 -1.1242  1.5688  0.1066  1.0018  1.6148
#>  1.3432 -0.5761 -0.4851 -2.3010  0.0231  0.0662 -1.2565  1.4197  0.3114  1.9180
#>  0.9678  1.2126 -1.1445  0.2386 -0.2716  0.1560 -0.7912 -1.3769 -0.1172 -0.1907
#> -1.6111 -0.2688 -0.4006 -0.3635  0.0229  1.5388 -0.0253 -1.7515  0.1212 -0.0610
#> -0.4675  0.6738  1.0080 -1.7725  2.2180 -1.9102  1.7393  0.7879  0.0885  0.2433
#> -1.3370 -1.0431 -0.5270  0.3307 -0.1436  0.0284 -0.2838  0.7333 -0.9288  0.7637
#> -1.7080 -1.4596 -0.0826 -0.4909  1.5067 -1.8714  0.4314 -1.3538  0.4415  0.5717
#> -0.0508  0.7288 -0.3699 -0.8834  0.5755 -1.9155 -1.2935  0.3502 -0.4930  0.0902
#> -0.5164 -0.7284 -0.1173 -0.4718  0.5347  1.1919  0.0261  1.1630  0.1744 -1.5644
#> -1.4434 -0.1834  0.4173 -0.0456  0.6473  3.2193  0.7359  0.9265  0.5109 -0.1682
#> -0.8362  0.1967 -0.6683  1.6057  0.7546 -0.8045 -0.3485 -1.0114 -1.1746 -1.8119
#> -0.0906 -1.3000  2.5013 -0.6431  1.3664 -0.2497  1.7314  0.4511  0.8209 -1.3381
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]