Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3037  0.7389 -0.3426  0.1947  0.6615  0.5293 -2.1114 -1.1831  0.2930 -2.0029
#>  0.2736 -1.3632  0.0257 -0.0292  1.1265 -0.3419 -0.3013  1.2240 -1.3822  1.3406
#>  0.3518  0.7268 -1.3157 -0.3030  0.7953 -0.0930  1.6674  0.1261 -0.9850 -0.1584
#>  0.1758 -1.3844 -2.2472  0.7271 -0.3460  1.2069 -0.5023 -1.4687 -0.3061  0.1934
#>  0.8489 -0.3817 -0.0137  1.2490 -1.2000 -0.3953 -0.5173 -0.7718  2.3540  0.4707
#>  0.4513  0.3529 -0.5742  3.0567  0.0568  0.3092  0.3037 -1.3690 -0.2272 -0.7980
#> -0.5886 -1.1142 -1.6091 -1.3107 -1.0186  0.4084  0.7853 -0.1798  0.5743 -0.6338
#> -0.1727  0.0894  0.4627 -0.3079  0.7411  0.1644 -1.3329  0.3935 -0.1359  0.2904
#>  1.6033 -0.3585 -0.2547  1.9061  0.1266  0.5710  1.0374 -1.9761  1.3051  0.1576
#> -1.4718  2.1797  0.1596  1.2242  0.6286  0.1394 -0.0921  1.3794 -0.8332 -0.5258
#>  1.9306  1.5238 -0.3366  0.0331  3.5419  0.4495  0.8592 -0.2753  1.3853 -0.6722
#>  0.1700  0.6170 -1.2694  0.9350  1.1790 -1.3529 -0.9681  0.3185 -1.4562  0.1867
#>  0.9222 -0.3777 -0.4565 -0.3847  0.7736  0.7717  1.3044  1.3689 -0.4082 -1.1031
#> -0.8171 -0.0968  0.4672 -1.7971 -0.5934 -1.7783 -0.7243 -0.0804  0.4923 -0.8421
#> -0.1542  0.6380 -1.0417 -0.3282 -0.5841 -0.8740  0.1383  0.5129 -1.4734  0.2548
#>  1.1782  0.3200 -0.4997  1.1853 -1.0122 -0.4842 -0.8611 -0.8109  0.3219 -0.6306
#> -0.9812  1.7248  2.0624  0.9558  1.2586  0.0727 -1.1029  1.8936  0.0888 -0.0382
#> -0.0734 -0.1639 -0.0409  0.0932  0.0481 -0.8170 -0.6333 -1.8233 -1.1345  0.3409
#>  0.6968 -1.2671 -1.3264  0.7552 -0.7194  2.3245  0.7102 -0.4863 -0.5600  0.4260
#> -0.8087  0.4504  0.4220 -0.1533  0.5269  0.0717 -1.9384 -1.1896  1.0988  0.4349
#> -1.4902  0.5450  0.1648  0.2810  1.6180 -0.5957 -1.1374 -0.2802  0.9364  1.5733
#> -1.9936  0.7931 -0.8717 -0.9621  1.6370 -0.1013  0.4175 -1.1495 -0.4568 -1.8576
#> -1.0380 -0.5444 -0.0257  1.4743 -0.9955  0.9713 -0.8041  0.2895 -0.3717 -0.3031
#> -0.3649  1.4269  0.8824  0.1209  0.3040 -0.7915 -0.2271 -1.0256  1.5106 -0.6269
#> -0.7591 -0.7704  0.8878  1.0818  0.1367  0.9866  0.9618  0.9288  2.2801 -0.6848
#>  1.3639  1.1815 -0.7561 -0.1481 -1.7463  0.5466  0.0494  0.3141 -0.0892  2.5094
#> -0.9085 -0.1678 -0.7353  0.0996 -0.3791 -1.2646  0.7887  0.0406  0.0370  0.6152
#>  2.5502  0.1653 -0.5199  1.1628 -0.4969  2.2382  0.6872 -0.2397  2.2792  0.3017
#>  0.4244  0.1578 -0.1115 -1.8925  0.5993 -0.2204  0.1952  0.9010  2.3321 -1.0454
#>  1.4827  0.7812 -0.5630 -0.8058 -2.5980 -0.2660  0.1734  0.5267  0.4055  0.6930
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]