Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.9104  0.4095  0.2122 -0.2090 -0.1665 -0.1751  0.2465 -2.0943  0.5893  0.1472
#> -0.7457 -0.3482 -1.1304 -0.0109  0.5144 -0.5403 -1.4190 -0.4862  0.8507  1.7210
#> -0.8107  0.2473 -1.9931 -0.2001  2.6386  0.6066  0.5474  0.2978 -0.9957 -0.5478
#>  0.7434  0.2206  1.2281 -0.7203 -1.3679 -1.4683 -1.4364  1.3403  0.9149 -0.0661
#>  1.5185 -0.9700 -0.9882 -0.4558  0.7091 -0.6953  0.4477  0.6545 -0.8693  0.1877
#>  0.4215  1.2979  1.2748  0.0701  0.3261  1.1088  1.9605  0.0296 -0.0311 -0.7023
#>  0.0255  0.0472  1.2680 -0.7565 -0.5245  0.2526  0.0232  2.1489 -0.2926  0.5538
#>  1.4778 -0.7445 -1.8377 -0.1139  0.6271 -0.6556  0.2593  0.6706 -1.2658 -2.8360
#>  0.3125 -0.3422  2.8740 -2.1017 -2.5020 -1.4473 -0.2052 -2.5944  0.7548 -2.9303
#> -1.1083  0.2095 -1.0943  0.7371 -0.0262  2.0062  1.1883  0.5685  0.2784  0.5683
#>  0.8784  0.6680  0.0364  0.6817 -0.4511 -1.0648 -0.9275  0.2225  2.1303 -3.8671
#> -0.1162  0.2101 -0.2132 -1.3314 -0.3115 -0.3325  0.7091 -1.2475  0.6520 -0.4150
#> -0.3551 -0.8443 -1.0342  0.7924 -0.5656  0.2442 -0.2356  0.8711  0.1331 -2.9048
#> -0.2991 -0.2253  0.0607 -0.4332 -1.7537 -0.1792 -0.0367  1.6995  1.0322  0.4321
#> -0.3282  0.7796  1.4007 -1.3327  0.6002 -1.1756 -1.4554 -0.2669  1.0806  1.2923
#>  0.3028  0.0406  0.8994 -0.1104  1.1312 -0.2444 -0.7481 -0.6515 -0.4050  1.0369
#> -1.6952  0.2242 -2.2977  1.0323  0.0304  0.3246 -0.3641 -1.1989  0.9689 -0.0144
#> -1.2344  1.2861  1.7340  0.9228  0.6623  1.1598  0.0545  0.8989  1.4185  0.0164
#>  1.1694 -0.4244 -0.7088 -1.3144  0.3093 -0.7610 -0.9549  0.2277  0.5824 -0.6292
#> -0.8522 -0.2276 -0.2028 -0.6064 -1.5260 -1.8387  2.4958  1.1343 -0.3775  0.4204
#> -1.4312  0.6409 -0.7153  1.1852 -1.0684 -1.4816 -1.1295  0.7781 -1.5156  1.3144
#> -1.1109 -1.3616  0.3467 -2.4289  1.7878 -1.1103 -0.7358 -0.3986  0.4777 -0.5473
#>  1.3597  0.9179 -0.3685 -0.8248  1.8864  1.2431  0.5832 -0.2930  0.6637 -0.2700
#> -0.6962 -1.4483 -1.7333  0.7065  1.3034 -0.3223 -0.3767  1.8671  0.2710  1.7828
#> -0.7523  0.5999 -0.1861 -0.5665  0.2959  0.2880 -2.8523 -0.3245 -1.0845  1.3781
#> -0.4833 -0.2381 -1.6311 -1.2482  1.2376 -1.3393  0.1889 -1.0628  0.3594 -0.4824
#> -1.1125  0.2234  0.5578  0.1234 -0.0962  1.8423 -0.2690 -0.4947 -0.3986 -0.7633
#> -1.6226 -0.1308  0.3086 -0.8569  0.2117  0.0310  0.3568 -1.9387 -0.4671  1.8489
#> -1.0852  0.4159 -0.8869  0.0792  0.1439  0.8709 -0.6255 -0.3090 -0.3371 -0.1371
#>  0.1743  0.3294 -0.4266  2.0407 -1.9789  0.1256  0.1867  1.8659  0.4804  1.1954
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]