Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.3479 -1.2559  0.3755 -0.2688  0.3789  0.6921  1.6401 -1.3321 -0.8218 -0.2992
#>  2.2820 -3.3397  0.5820  1.3297  0.5412  1.8974  0.3830  0.5236 -1.6910 -1.5749
#> -1.3627  0.0950 -1.0257 -1.5193  0.5036  0.5340 -0.0724 -0.1370 -0.2462  0.0900
#> -0.4837  0.1363  0.8848  0.2232 -1.1231 -0.9454  0.8383 -1.9312 -0.1202  0.5433
#>  0.6280 -0.5092 -0.5503 -1.0308 -0.2230 -0.6739  1.2708  0.2335  1.4842  0.3954
#>  1.2827  0.0026 -1.0569 -1.5158 -0.4385 -0.7052  1.2244 -1.7065 -0.3417  0.1228
#>  0.5588 -1.2972  0.5772 -0.2864  0.0289  0.6635 -0.4683  0.8340 -1.5312 -1.8499
#>  0.0004 -0.2054  0.8701 -0.0056 -0.6239 -0.5557  2.3189  0.4228  2.3873  1.7424
#> -0.3392  1.0830 -0.4504  0.7752 -0.1372 -1.0516  1.2979 -1.9941  1.6142  0.4719
#> -0.2122  1.5878  0.3685 -0.7585 -0.0720 -1.1340  1.0813  0.1655 -2.7219  0.6006
#> -3.2246 -1.3404  0.5418  2.1103 -0.9038  0.2381  3.0098  0.1473 -1.1774 -0.3651
#> -1.3241 -2.5420 -0.9228 -2.1887  0.4972  0.0560  1.4211 -0.6641  0.6975 -0.9337
#>  1.3528  0.3758 -0.4821  0.6641 -0.5465 -0.4601 -0.0366  0.6438  0.0895 -0.4968
#> -0.2775 -0.6275 -0.3560 -1.5190 -0.7593  0.2381  0.8684  0.3457  0.7689 -0.8022
#>  0.5660 -1.4602  1.0130 -0.8048  1.1681 -0.2642  1.2055 -1.2838 -0.5415 -0.0629
#>  0.7974 -1.2944  0.1927  2.3236 -1.1896  1.1110  0.9783  0.8142  0.0606 -1.6780
#> -0.4199 -0.2509 -0.6325  0.3122  0.0842  0.7556  1.4466  0.7251 -0.2393 -0.3604
#> -0.7882 -0.3145 -0.0653 -0.8219  0.7703  0.7465  0.0049 -1.1352 -0.0289 -1.0448
#>  1.5008 -0.8812  1.2224 -1.2473  1.9732  1.0207  0.6907 -1.0257  0.3568  0.4390
#>  0.3588  0.5856 -0.4952 -0.0584 -1.3070  0.5832  0.5087 -1.0875  0.9184 -0.0330
#>  1.9765 -0.1993 -0.0320 -0.5126  0.9588 -0.0465 -1.5412 -0.0769  0.9674 -0.2141
#> -1.2399 -0.1433  0.6730 -0.8412  2.0650  0.6279 -1.8648 -0.8164 -0.2442 -0.5737
#>  0.7061 -0.8678  0.5926  0.2913 -0.7191 -1.3439 -0.0904  0.4048  1.0224 -1.5662
#>  0.2968  0.9225  0.0923 -1.1795  1.7404  0.7744 -1.2712  1.6685 -1.0009  0.7886
#> -0.6851 -0.8517  0.2729  1.1764 -0.2653 -0.8263  0.1849 -0.5703 -1.4831  1.7466
#>  2.0476 -0.9202 -1.4270  0.8896 -0.4311 -0.8143 -2.0646  1.1272  1.0481 -1.0642
#> -0.3016 -0.4698  0.2437 -0.8839 -0.8708  0.7819 -1.1467 -0.6021 -0.6984  0.1614
#> -0.8551  0.6317 -0.4136 -0.0393  0.4116 -0.7594  0.4023  1.5162  0.2501 -0.3428
#> -0.3332  1.1805 -0.5204  0.3301 -0.1380 -0.2453 -0.0115 -2.5115 -0.3495  0.4494
#> -0.3133 -0.5261 -1.6519  0.3152 -1.6325 -1.3942 -0.6351  1.9305  1.0895  1.5406
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]