Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.8540 -1.8206  0.2810  0.4478 -0.9570 -0.4631 -0.7458 -1.9362 -2.0406  0.1044
#>  1.0986 -0.5790 -0.6050 -0.7762  0.5953  0.1038  0.0814 -2.7806  0.8263 -0.9954
#>  0.5023  0.7308 -0.3163  0.5901 -0.2150 -0.8374  0.4834  0.2053  0.5706  0.3404
#> -0.0112  0.3162 -0.6511 -0.4464  0.8506  0.4073 -0.8120  0.4339  2.6389  0.3541
#> -2.6615  1.1696  0.9258 -0.5243  0.4770 -0.5568  0.0535 -0.8365  0.2531  1.1417
#>  1.0802 -0.3516 -0.1653 -0.7506 -1.6687 -1.7307 -1.9167 -0.5701  0.4244 -0.3654
#>  1.5551  0.5042  0.2604 -0.5384 -0.0098  0.0686  0.1891  0.4035 -0.2905  0.1956
#> -1.8654 -0.9320  0.3213 -0.4378 -1.1158 -0.5368  0.1253  2.2867  1.1013  0.1970
#>  1.1237  1.1690 -0.3154  1.0847  1.8336  2.4945 -0.2112  0.7243  1.8161 -0.7196
#>  1.0459 -0.2152  1.0022  0.6037  0.4012 -0.2919  0.2873 -0.7662  0.3341  0.1120
#>  1.0055  0.0449 -1.6374 -0.4004  0.3846  0.1239  0.0186  0.6227  0.2092 -1.8123
#> -0.3805 -0.4971 -0.5849 -0.0978  0.0593  0.6956  0.5768 -0.3565  0.1492 -1.8609
#> -1.1157 -1.1210  0.2535 -0.5458 -0.4931  0.5626 -0.5186 -1.5350 -1.3798 -0.3417
#> -1.2595  0.0111 -0.0531 -1.9273 -0.0311  0.1067 -0.1619 -0.2695 -1.1565  0.8895
#>  0.4381  0.6222 -0.6553  0.7058 -1.6402  0.2717 -1.7702 -0.6482 -1.5315 -0.8208
#>  1.1610 -0.2282  1.8263 -0.6634 -0.7440  0.2628  1.0886 -0.1274 -0.4874  1.2419
#> -1.7706 -0.1576 -0.2509  1.0820 -0.4440  2.2775  0.0200 -0.5096  0.1602 -0.5212
#>  0.2386 -0.6027 -1.6440  1.0510  2.9396  0.3500  2.0585  0.0004  1.2897 -0.0682
#>  0.6657  0.0445 -0.3075 -0.3853 -1.5041 -3.6333  0.1469 -0.7558  2.9920 -1.0057
#> -0.6492 -0.0188  0.3816  1.8191  0.9901  1.1478 -2.1586  0.1680 -1.9403  1.7709
#>  0.2749 -0.0503  0.8148 -1.8568 -1.3907  1.0567 -0.7801 -0.0445 -0.3503  1.0017
#>  1.3963 -0.1471  1.6610 -1.5021 -1.0753 -0.6095  0.4982  1.1089  2.2669 -0.5359
#> -0.2511 -0.0255  0.7649 -0.2347 -0.5285  2.0019  0.0973 -1.4908 -0.2678 -0.0230
#> -2.5105 -0.0857  0.7360  0.3618  1.6351  0.2497 -0.7668  0.5245  0.1959 -1.1796
#> -0.8463  0.2481 -0.2175  1.6078  0.6532  0.6553 -1.5685 -0.8218 -1.5990 -0.1972
#>  0.6803  0.0010  1.0140 -0.4194 -1.3041  0.9937  0.2158 -1.3085  1.0711 -1.4334
#>  1.0746 -0.8651 -1.3175  1.3602 -1.5351 -0.0271 -0.3188 -0.8122  0.2510 -1.1800
#> -0.8205 -0.2818 -0.9700 -0.4990  0.7901 -0.5585  0.2547 -0.8613 -0.7074 -1.3060
#> -0.1934 -0.2729  1.7335  2.1808  0.2578  0.0271 -0.7074 -0.6494  0.3894 -0.4615
#> -1.3065  0.1641 -0.3008 -1.2306 -0.6138  1.8754  0.3411 -1.5805  1.1401 -0.6033
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]