Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3664  0.3055  0.0368  0.6687 -0.0054  0.5449  1.4932 -1.0634  0.1751  0.9644
#>  0.4534 -0.0506  1.7974 -1.5879 -2.7556 -1.3238 -2.5788 -0.8574  1.7640 -0.5773
#>  0.1751  0.6839 -0.2892 -0.5769 -0.1582 -0.4002 -0.6222 -0.8089 -0.0469  0.0887
#>  0.6915 -1.3585  0.8819 -1.2504  1.8579 -1.1489  2.1565 -0.0277  0.5546  0.4541
#> -1.4589 -2.0105 -1.9165  1.7193  0.3535 -2.4135  0.3444 -0.7372  0.3993  0.7343
#> -0.6523 -1.3356 -0.4130 -0.4134 -0.2254 -0.4231 -0.5162 -0.0808 -0.8187 -1.0443
#>  0.5765  0.2826  0.3401  1.4152 -0.3865  0.4610  0.4173  1.7051 -0.8458 -0.3074
#>  2.0013  0.7558 -0.2668 -1.4023 -0.9524 -1.0110 -1.3906  0.1781  0.4019  1.6730
#>  0.2085 -0.7099 -0.8298  0.0766  0.1566 -1.5705 -0.3210  2.0980  1.2540  0.0355
#> -0.8327  0.8637 -0.1642 -1.1303  1.4904 -0.7455 -0.0820 -0.9259  2.4284  0.9448
#>  0.7414  0.1576  0.3882 -0.1775  0.7167 -0.4352  0.3978 -1.1750 -0.6630 -1.7074
#> -0.3117 -0.9549  0.3469  1.0100 -1.5847  0.2908 -0.1241  0.1132  0.4926 -0.6555
#> -0.4425 -0.9971  0.1489 -1.3803 -1.4877  1.0182 -0.7188 -1.5673 -0.4623  0.0774
#>  0.0417  2.0470  1.2472  0.4991  0.8808  1.0650  1.6359 -0.1710 -0.5205 -0.0746
#>  1.5805 -0.4091 -0.7797 -0.4399 -1.0807  0.5867  0.3243 -0.1975  1.2806  0.8712
#>  0.6397 -0.4028 -0.6940  0.5208  1.2493  0.2630  0.4461 -2.2427  0.0502 -1.1681
#> -0.2355  0.9292  0.6911 -1.7903  0.8671  1.6012  0.1374  0.1547  1.4370 -0.5953
#>  0.7968  0.8743 -0.9049  0.7539  0.3111 -0.9711 -0.9241  1.4820  1.2280  0.9117
#> -2.6848 -3.0374  0.0894  1.4361  0.2323 -0.3112 -0.0385 -0.6166  0.5268  1.0792
#>  0.6060 -1.8016  0.4647 -0.5362 -0.4278 -1.9315  1.0952  0.5818 -0.5033  0.1696
#> -1.3347  0.2875  0.4309 -0.1422 -0.5090 -0.5945  0.0440 -0.2767 -0.8641  2.2493
#>  1.1958 -0.0872  0.9373 -1.0552 -1.0281 -1.3589 -2.3002  0.4946  0.1313 -0.9264
#> -1.2588 -0.4122 -0.4802 -0.2037  0.3922  0.1219 -0.0299 -0.7171  0.1830 -1.2924
#> -0.6533  0.2814 -0.7478  1.1689  0.2974  1.5081  0.5483 -1.9982 -1.5245 -0.4196
#>  1.5432  0.6871 -1.8576  0.2413 -1.4775  2.4116  1.2274  0.6476 -0.0193  0.2135
#>  1.5565 -0.0548 -0.6868 -0.1176  0.5856  0.3961 -1.7994  1.2228  0.9405 -1.1533
#>  0.8636  0.1518 -0.5888  1.2353  0.2483 -1.8197 -2.2685  2.4077  0.0256  0.3071
#>  0.6132  1.6527  0.9543  0.6393 -0.5803 -0.6913 -0.8669  1.5774  0.5191  0.8505
#>  0.1161  0.4530 -1.0727  0.7519  0.4930  0.1326  1.8463 -1.5683  0.0092  1.8702
#>  1.1146 -0.1602  0.6296  0.8446 -0.4111  0.8611  0.4777 -0.2103 -1.6283 -0.7392
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]