Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.4418 -1.0232  0.5652 -0.6725  1.0707  0.8243 -0.5029  1.1369  1.7844  1.7787
#>  1.0244  1.4671  1.5992 -0.1252 -0.2029 -0.2322 -0.7449  0.8718  0.7637  0.7218
#> -0.2338 -0.4254  0.0013 -0.6306 -1.2245 -0.9903 -0.8480  0.4515 -0.2033  1.0895
#> -0.9419  2.1820 -0.4921  0.4745  0.6637  0.9859  0.1319 -0.0522 -0.0929 -0.4946
#> -0.4946 -1.6786 -0.2039 -0.0384  0.6565 -2.5066  1.5931  0.5787  1.0484 -0.1424
#>  0.4263 -0.4639 -1.3513 -0.5796 -1.0183 -0.6482 -2.0648  0.4536  0.5613  1.6690
#>  1.7113  1.7696 -0.8010  1.7828  0.6785 -1.1077 -1.4016 -0.4456  0.8925  0.5182
#> -1.5709 -0.1144  0.5119  1.1206 -0.8786  0.5214  0.1609  0.7605 -1.8791 -2.2492
#> -0.3668 -1.7254 -1.0111 -0.0123 -0.7815  0.1914  0.0841 -0.0342  0.7648  1.5115
#>  0.9594 -0.0422  0.0826  0.9370  0.6449 -1.2252 -1.6248 -1.6515 -0.4933  0.6985
#> -0.6096 -0.2779 -0.1385  0.9562 -0.6388  0.4306  1.1313  0.7602  0.0973  0.2228
#>  2.5226 -0.8947 -0.4505  0.6514  1.2403  1.4039  0.5387 -0.2161 -1.5079 -0.7425
#> -0.3333  1.3605 -0.4932  1.0230  2.3371  1.3513  0.2894 -0.0788 -0.3492  0.9147
#>  0.1984 -0.8034  0.2678  0.1469 -0.2190  1.9114  0.1148  2.0833 -0.0911  0.5453
#>  0.5875  0.8549 -1.6480  0.2253  0.8209 -0.3295  1.6194  1.1556  1.0154 -0.2878
#> -1.0152  0.8790  0.6445 -1.7995  0.9674 -0.9922  0.5527  1.6972  2.3151  1.6664
#>  1.0124 -1.8290  1.1472  0.6623  0.0876  0.9658  1.0727 -0.3758  1.2015 -0.3192
#>  0.2747  0.2393 -0.6375  0.9295  2.8400 -0.4720  1.0444 -0.2277 -1.9873 -0.1113
#> -1.0545  0.0855  0.0481 -0.0969  1.1167  1.0514  1.3441  2.2359  0.2177  0.3779
#> -0.3886 -1.5803  0.1740 -1.1550  0.1549 -0.1367 -0.3797 -0.4666  0.3111  1.0127
#> -0.0581  0.2021 -0.2982 -0.2636  0.8800  1.0892 -1.9065  1.4596 -0.4643 -1.7464
#>  0.3567 -0.9981  0.5554 -0.1356 -1.0439  1.1089 -1.5192  0.9677  0.5960 -0.5768
#> -0.5486  0.3788  0.4102  1.1441  0.1579  0.5627 -1.6388 -0.4764  0.4351 -1.5138
#>  0.7498  0.5904  1.5714 -0.0326  1.0201 -1.1707  0.2520 -1.5309  0.8408 -1.8192
#>  0.6297  2.2692  1.2962 -0.6875 -1.5231 -0.2257  0.9128 -0.7030  0.6271 -0.2124
#>  0.5614 -0.7771 -0.5928 -0.2966 -0.4865  0.4924 -1.1285 -0.3259 -0.9513  0.9167
#> -0.8018 -0.9706 -0.0167  0.9918  0.2068  0.8551 -0.7341  1.1745 -0.5055  0.4341
#> -0.2842 -0.6304 -0.5939 -0.7677 -2.2367  0.3251  1.4555 -0.3383  1.2063 -0.2821
#> -0.8867 -0.4848  0.6135 -0.4842  0.6642 -0.5844 -0.3779 -1.1833 -0.7265 -0.6521
#>  0.1642 -0.4502  0.7362 -0.2635 -0.5479  1.3018 -1.7105  0.1178 -0.9298 -0.5290
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]