Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.4440 -1.4239 -0.9020  1.1929  0.7937  1.5687  0.1414 -0.4780 -0.0403  0.0812
#> -0.4012  1.7735  0.5311 -0.1986  0.1781 -0.9768  0.2074  0.6425 -1.1809  0.0429
#>  1.1376  0.8429 -0.8856 -1.4149  0.0383 -0.2742 -1.5587  0.6788 -1.8359  0.0729
#>  0.7031 -0.6114  0.4198  1.2913 -1.0304  0.6213  0.4115  0.2965 -0.2807 -0.3484
#> -0.1922 -1.7984  0.2129 -0.2991 -0.0735  0.0013  1.0842  0.6751 -1.4409 -0.6170
#> -0.7931 -0.3537  1.0356  1.3599  2.5364  1.5748  0.7744  0.3485 -0.1315 -1.0733
#>  0.0869  0.0057 -0.2351  0.3384 -1.0330 -0.7111 -0.7366  0.9124  0.2048 -0.3552
#>  0.3654  0.3027 -0.7569  0.8716  0.7001 -1.1190 -1.2425 -2.2530 -0.0570 -0.6715
#> -0.3836 -0.0867 -0.5940 -0.5404 -0.0077  0.0523 -0.6923 -0.2886  0.0059 -1.1016
#> -1.4504 -1.9790 -0.0944 -0.1273 -0.2808 -0.1898  1.4121  1.8772  0.5212  0.1107
#>  1.0112 -1.3709 -0.3534  0.0060  0.7808  0.0429  0.6332 -1.3424  0.5885 -1.0293
#> -1.7090 -0.7998 -0.7486 -1.1947  0.1599 -3.0930 -0.3255  0.0889 -0.6354 -0.2682
#>  1.1881 -0.1672 -0.0941 -0.6455  0.9138 -0.7499  0.7529  0.8413 -1.0886  0.3501
#>  1.4717 -0.8036 -0.4152  0.0558 -0.6525 -2.1338  0.2096  0.5633  1.3274 -0.0901
#> -2.4385  1.0661 -1.7500  0.2111  0.0443 -0.1089 -1.2999 -1.3135 -0.6463  0.6860
#>  0.1433  2.2829 -1.1566 -0.9014 -0.3741  0.4491 -0.8227  1.5885 -0.0273 -0.2913
#> -0.1883  0.7407 -0.7275 -0.0016  0.7056 -0.7377 -0.8420  1.4010  0.1611 -0.5683
#> -0.8473  0.8420 -0.1574  0.8950 -2.3046  1.0112  1.4446 -0.4525  0.4665 -1.1262
#> -0.5370  1.2203 -1.1089 -0.6410 -0.8180 -1.6778  0.6970 -0.5299  0.0566  1.8193
#> -0.5687  0.1073 -0.2735  0.5436 -2.2574 -0.6727  0.0557 -1.6742  0.6548  0.4869
#> -0.9304  0.4894 -0.2248 -0.2758 -0.2427  0.5128  0.1748  0.9383 -0.3314  0.9643
#>  1.5331  1.4478 -0.2215 -1.7943 -1.7378  1.7856  0.8418 -0.1902  0.4056  0.0164
#> -0.0657 -0.1709 -0.0863 -0.2613 -1.4589 -1.4329  1.6938  0.2432 -0.5069 -1.7430
#> -1.1279 -0.9705 -0.8044 -0.7294  0.6677  1.2234  0.9898 -1.0151 -0.5044 -1.2813
#>  0.3404  0.0489 -0.6237  0.7677  0.3875  1.3256 -0.8339  0.2230 -1.2627  1.6871
#> -0.2311  0.8006  1.1752 -0.1859 -0.1359 -0.0332  1.1441 -0.5868 -0.0045  1.1958
#>  0.5812 -0.8932  0.7437  0.4859 -0.8476 -0.9696  0.4072  0.3167 -2.3075  0.8084
#>  1.7340 -0.7081  0.4187  0.6259 -0.4348  0.2714 -0.5964 -1.1316 -0.2006  0.1416
#>  1.8073  2.7300  0.4183  0.4785 -0.2312  1.4268 -1.2426 -1.8226  1.1263  0.1533
#>  0.1264 -0.0686  0.0038 -1.1565  1.2538  0.7705  0.0735  0.3506 -2.0079  0.8043
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]