Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0927 -0.7452 -1.2150  0.1963 -0.6881  1.5230  0.2829  0.6666 -0.6063  0.0542
#>  0.5405  0.7612 -0.4709  0.9869 -1.2692 -0.8239  1.5001 -0.6652  0.2228 -0.8727
#>  1.0419  2.1995  0.3294  2.7549  0.1195 -0.2469 -1.7052  0.2013  0.1206 -0.9933
#>  0.3548  0.8848  1.5103 -0.8708 -0.9247 -0.4852  0.5161 -0.5960 -0.3819 -1.3428
#> -0.8496  1.3955 -0.4324 -1.7552 -0.9882  1.5571  0.5371  0.3818 -0.8607  0.7465
#> -2.0885 -1.3564 -0.6002 -0.0271 -0.3347  1.1699  0.7289 -0.3711 -0.2174  0.3563
#> -1.4654  0.8592 -0.1199  0.4509 -0.0255 -0.3386  1.8664  0.2017 -0.6111  0.2864
#>  1.5617  0.4020 -0.2336 -1.6602  0.0349 -1.4967  1.3596  0.5766 -0.3302 -1.3087
#>  0.8275 -1.3674  0.1306 -1.0910  0.3333 -2.0323 -1.2535 -1.2497 -0.9980 -1.1301
#>  1.3338 -0.7095  1.1226  1.2206  0.4328 -0.2681  1.2004 -0.7262  1.1132  1.4897
#> -1.1310 -0.5351 -1.6305  0.7271 -0.7529 -0.1804  1.8597  0.5291  0.4758 -0.4733
#> -0.1696 -1.0115  0.6087  0.3085  1.0699  0.2704 -0.6424  0.5407 -0.4796  0.5488
#>  1.0702  0.4270  1.0550  1.7589  0.6218 -0.5430 -1.4357  0.8546 -0.3917 -0.8347
#>  0.4293 -1.0788  1.0903  0.6936 -0.2990  0.4587  1.2930  0.0974  0.0178  1.0732
#>  0.8756 -0.0546 -1.2050  0.2866  0.4614  1.1421 -1.1945  0.2114  0.6763 -0.1132
#>  0.4986  0.2656 -1.0711  0.1151  1.4503  0.8829 -0.5647  1.9646  0.7253 -1.7758
#>  0.1972  1.0088 -0.1062 -0.1579  0.9722 -0.5851 -1.7345 -1.3243 -0.5404 -1.2106
#> -2.0577 -0.6608  0.5964  0.3818  0.1515 -0.1277 -1.0230  0.0605 -0.5442 -0.2176
#>  0.1587 -0.9636  1.3594  0.6291 -1.4834 -0.0961 -1.1704  0.8344  0.2966 -1.2002
#>  0.4395  0.3314 -0.1644  0.3554  0.5972 -1.9066 -0.6204  0.0623  0.2521  0.5183
#>  0.5114  1.1942  0.3728 -0.2820 -2.8886  0.3425  1.4984  0.1001  1.4486  1.3198
#> -0.6411  0.5834  1.4711  0.4057 -2.0042 -1.4348  0.5668 -1.5372  0.5924  0.0500
#> -0.2311  0.4746  0.6400 -0.2306  0.3776  1.1377 -0.6040 -0.7712 -0.5622 -0.6402
#> -1.4021  0.6191  0.8097 -0.9112  0.3997  0.9646 -0.7491  1.0930  1.4468 -0.3237
#>  0.8750  0.6264 -1.1316 -1.1854 -0.5847  1.3260  1.9505 -1.1218  1.3768  0.0389
#> -1.3594  0.3161  1.0571  0.0204  1.6597 -2.2610 -1.0952  1.1558  1.4823 -2.7005
#> -0.2872  0.6802 -0.6390 -0.0505 -0.3091 -0.3019  1.2369  0.4734 -1.4547  0.8213
#>  0.0517  0.2751 -0.2748  0.3152  1.8562  0.6298 -2.2922  0.3085 -0.3193 -1.6708
#>  0.2366 -0.0976 -1.0526 -2.9939 -0.5067 -0.1758 -0.0503  0.3319  1.8461 -2.2055
#>  2.2309 -0.3227 -0.0830 -1.0225  0.6831 -1.2754  0.1004 -0.5234 -1.1228 -1.4741
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]