Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7678  0.6950 -1.1516  2.3181 -1.3003 -1.0979  0.6923  0.7106 -0.3745 -0.5477
#>  0.4677 -0.4049 -1.3425  0.7024  0.3849  0.2277 -1.6140 -0.4369 -0.3198 -1.0243
#> -0.7530 -0.5710  0.0628  1.0854 -0.5329 -0.6839  0.6205 -1.9945 -0.9666 -0.9170
#> -1.3201  0.0914 -0.3925  0.8913  0.8525 -0.5428  1.2635  1.7443  1.7574  0.5014
#>  0.0708  0.0876 -0.7212  1.2957  0.3696  0.7821  0.5105 -1.3576  0.8174  1.3152
#> -0.5159 -0.2949 -0.8708  1.5908 -0.8540  0.7076 -1.1057  0.3189  1.2892 -1.1622
#>  0.6415  0.9209 -0.9844  0.8874 -1.0566 -0.6945 -1.6825 -1.1806  0.2046 -1.0000
#> -0.2108 -0.2825 -0.0834  1.3104 -0.5180  0.9218  0.6206 -0.6199  1.1669  0.8468
#>  0.7472 -0.6539  0.3601  0.4034 -1.0635 -0.5459  1.0672  0.9037 -1.5055 -0.9817
#>  0.4492  0.3117 -0.6395  0.3866 -0.2426  0.6890 -1.4610 -1.6193 -0.7799  0.1667
#>  0.7318 -0.6750 -0.2429  0.0440  1.0110 -0.2443  0.5290  0.6265  0.6298 -0.1840
#> -1.6087  1.5534  0.6964  1.1056  0.6501 -0.2309 -0.1066 -0.3278 -0.5506 -0.0602
#>  0.0733  0.5868 -0.9125  0.6793  0.6286  2.5380 -0.7660  1.6031 -0.1260 -0.1166
#> -1.3399  0.0286 -0.2542 -0.6964 -0.3086 -0.8793 -1.0383 -0.1044 -0.2834 -0.1979
#>  1.2453 -1.1188 -0.3730 -1.4688  0.9726 -2.9303 -0.5333  0.0461 -0.4479  0.6662
#> -0.8989  1.8872  0.3347  0.7662 -0.3431 -0.1045  0.6374  0.8438  0.7993  1.3124
#> -0.6396  0.3594 -0.7465  0.2244 -1.5118 -0.7212 -0.8319  0.5590 -1.0025  1.8870
#> -0.0071  1.1664 -0.1885 -0.4494  1.3811 -1.2885  0.2675  0.1819 -0.1901 -0.8120
#>  0.5605 -0.5190  1.2276 -0.6642  0.0505 -1.2296 -0.7425  1.6483  1.5442  0.8268
#> -0.5629  1.1781  0.0726 -2.3939 -0.0115  0.4684 -1.1777  0.2675  0.7708  0.9826
#>  1.9366  0.3261  0.3713  0.0980  0.3039 -1.1656 -0.2393  1.2496  0.6186  0.9573
#> -1.2710 -1.8651 -0.3866  1.1181 -0.5086 -1.7061 -0.2169 -0.4359  0.6947  1.4825
#> -1.0284  0.6183  0.5685 -0.0814  0.7933 -0.4189  0.2524  0.2527  0.0464  0.3128
#> -2.8894  0.0311 -0.2043  2.0856  0.4684 -0.7146 -0.7786 -0.1369  0.1652 -0.8901
#>  0.2596 -0.8487 -0.2499 -0.8299  1.0038 -0.0466  0.3889  0.7076 -1.3134  0.8985
#>  1.5318  0.2542  0.5942 -0.8547 -1.8653 -0.2264  0.6368  1.0607  0.8291  0.6234
#>  1.7813  0.9638 -0.7438 -1.0791  1.6296  0.8604 -0.3983  0.4372  0.0538 -0.3883
#> -1.1879 -0.0138 -0.4019  0.1164  0.1167  0.4102 -2.2875 -0.4267  0.3825 -0.6851
#>  0.8294 -1.3342 -1.6241  0.6377  0.1141 -0.6913  0.3363  0.9014 -0.0110 -0.4529
#> -0.6464 -0.6372  1.1898  0.0340  0.4440  0.1462 -0.5826  1.0036 -0.9993  0.2587
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]