Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4795 -1.9006  0.8331  0.9250 -0.6725 -0.1360  0.2462  0.0036  0.9812  0.5019
#>  0.7114 -0.7959 -0.0327 -0.9530 -0.4094  0.4412  0.3553 -0.0129 -1.3511  1.7267
#>  0.7793 -0.1861  0.6072  0.9473  1.3020  1.2202 -1.8276 -0.2502 -0.1890  0.7203
#>  0.5500  0.1143  0.1751  0.7504  0.1150  0.7612 -0.8271 -0.6973  0.8551  1.3065
#> -0.1122 -0.2885  1.4047 -0.3462  0.9036 -0.4205  0.4540 -0.1992 -0.0610 -0.4212
#> -0.9488 -0.4636 -0.3080  1.1519  0.5341  0.3380  1.2719  0.4070 -1.5207  0.8706
#>  0.2630  1.8413  0.7419  0.6955  0.6010 -0.3473  1.4665  0.3327  1.5175 -1.4490
#> -0.1428  1.8645  2.5729 -1.4074  1.1058  1.7112 -0.1175 -0.0732 -0.1542  0.2591
#> -0.9658  0.0740  0.5242 -0.0677 -1.0335  0.6632  2.0193  0.7588  2.0943 -1.4768
#> -0.3664 -0.4583  0.6878  0.0867 -0.4089 -1.0990 -0.6637  1.1194  2.2621  0.5294
#>  0.7287 -1.0004 -0.7259 -0.4984  0.2723 -1.1615  1.6926  2.2167  1.1830 -0.2262
#>  0.5261  0.2465  0.9260 -0.9655  0.1646 -1.6305 -0.0842  0.3886  0.7077  1.6297
#>  2.6539 -0.8652 -0.7972 -1.0112  1.5643  0.9345  1.4107 -0.2472  0.5025 -0.0844
#>  0.3467  1.2506 -0.2605 -1.2294 -1.1241 -0.9557 -0.5607 -0.8799  0.5759 -0.2678
#> -0.1410  0.0766  0.6971  0.8435  0.1498  0.4644  0.5009 -0.6184 -0.6545  1.6236
#> -0.2998  0.1898 -0.7921 -0.3676 -0.5198  0.6265 -0.1153 -0.8735 -0.2428  1.0634
#> -0.1334  0.5709 -1.2149 -0.0809 -0.9699 -1.0404 -0.0258  0.4991  0.1774  0.1214
#>  0.3165  1.7430 -0.4947 -0.0048  0.4298 -1.6786 -0.7545  2.5139 -0.8167 -0.5163
#> -0.5886  0.7630 -0.0815  0.4437 -1.8511 -0.1943 -0.5465 -1.1360 -1.0459  0.1081
#>  0.8991 -0.3891 -0.0469  0.5715 -1.7904  1.2460 -2.0855  0.6761  0.1595  2.0810
#> -1.0429  0.2923 -0.2088  0.9389 -1.7584  1.1265  0.7862  2.0163  1.0621  0.6684
#>  0.4075  0.8893  1.1659  1.2009  0.9481  0.5087  0.6429 -0.6834 -0.0149 -1.0008
#> -0.2492  1.2134  1.4333 -2.1614 -1.2764  0.8978  0.2880  0.4011 -0.7344 -0.5148
#> -0.8656 -0.5358 -0.2713  0.4200 -0.8636 -2.6834  1.1481 -0.4560 -1.0180 -0.5270
#> -0.7851 -0.9744  0.2301 -0.0461 -0.4302 -2.4816  0.4751 -0.4157 -0.4629 -1.1848
#>  0.3439 -1.7849  0.2665 -1.2182 -0.5203  0.2473  0.5529  0.8710 -1.3057 -0.5915
#>  1.3077 -0.2151  0.6111 -0.4177 -0.6410 -0.0716  0.6484  0.6461 -1.6619 -0.4862
#>  0.3869  1.4108 -0.3505 -1.2869 -0.4310  0.8770 -1.3568 -0.4763  0.3388  1.3136
#>  0.0728  2.2950  0.4217 -1.1346  0.4270  0.2473  0.3913 -0.3909  0.5527  0.9012
#> -0.1660 -0.9438  0.9476 -0.3020 -1.0573  0.3029  0.9149  0.5450  0.3002  0.3603
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]