Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5915 -0.0910 -1.3278  0.5798  1.4570  0.3378  1.5214 -0.1664 -0.7035 -0.1727
#>  1.2395  2.0883 -0.5267  0.5108  0.1399 -0.7817  0.7106 -0.8336  0.4107  0.4213
#> -0.4596 -0.2934  0.2333  0.3470 -0.7390  0.1413  0.6564  0.0917 -0.7332  0.9021
#> -0.6992 -1.9109 -0.2212  0.6077  0.5287 -0.2143  0.5812  1.3439 -0.7845 -1.5411
#>  0.8345  0.4262  2.3388  1.0337 -1.5616 -1.0592  0.8354 -0.5543  0.7680 -0.5529
#> -1.3521 -2.3064 -0.0988  0.9668 -0.7187  1.0140  1.6096  0.0277 -0.4351  1.1931
#>  0.6254 -0.0378 -0.1485  0.3652 -1.0631  0.3558 -1.8749 -0.7993  1.1083  0.6010
#> -0.5829 -0.6952  0.0518 -0.3158 -0.4982 -1.1236 -0.2214  0.2850  0.7832 -0.0756
#> -0.9205  0.7757 -0.0585 -0.4281  0.6270  1.2455 -1.7490 -0.6538  1.6561  0.3965
#> -0.7234  0.1620 -0.5695 -0.2780  1.7850 -1.4979  0.9728  1.0977 -0.6790  0.7707
#>  0.6972  0.2240  0.1313  1.6362  0.7449  0.9865  0.5280  0.8511  0.6414  0.2369
#>  1.0517  0.8338  1.0983  0.0691 -0.1388 -0.2124 -0.9598  1.3660  0.1350 -0.4096
#>  0.8915 -0.8332  0.4539 -0.7179 -1.0558  0.3655  0.6026  0.0880 -0.7942 -1.4104
#>  0.1563 -0.9238  0.5496 -0.0799  0.2782 -2.2798 -0.8688 -0.0986  1.7322  0.5214
#> -0.3046  0.9591 -2.4729 -0.9124 -0.9781  0.1077  0.8567 -0.3696  0.4337 -0.3430
#> -1.2472  0.4284  0.2394  0.8719 -0.9889 -0.8669 -2.5171  0.7764  1.9596  1.3304
#> -0.6374 -0.5395  0.5184 -0.6467 -0.2284  1.1694 -0.0938 -0.7133 -1.4222 -1.8829
#> -1.8399 -1.3670  0.6700 -0.1815  1.1683  1.3411 -0.3915  1.4795  0.2892 -2.3540
#>  1.0043 -1.9657  1.0212 -0.1969 -0.7259  1.3853 -0.7325 -0.5055  1.0847 -0.8345
#>  0.2642 -0.5977  1.3775  0.5400 -1.4654  1.2365  0.3143 -0.2812  1.3049 -1.3267
#> -0.0039  0.0792  0.9302  0.7614 -0.7062  0.4074 -0.6790 -0.4821 -0.1529  0.0927
#> -2.2264  0.2836 -0.7833 -0.9623  1.1971 -0.0367  0.2762  0.6269  0.4019  0.6167
#>  1.5782  0.1666  1.6908 -0.2711  0.1371  0.1091 -1.9416 -1.2931  1.4713  0.0823
#> -1.2489 -0.3859  0.5222  0.5781  0.9794 -0.1699 -0.5400 -0.0766  0.1445 -0.8454
#>  0.7955  0.6356 -0.2872 -1.1821  0.2731  0.4952  0.1049 -0.7528  0.2210 -0.0043
#> -0.4298  0.8799  0.0466 -0.5307 -2.0621 -1.4572 -0.3564 -1.4868 -1.2752 -1.3012
#>  0.8499 -0.7676 -1.8199 -1.0146  0.6172  2.0840  0.1719  1.5961  1.2091 -0.4489
#> -0.6294 -0.1353 -0.5099 -0.8948  0.9332 -0.5691 -0.2535 -0.2665  1.5962 -0.1034
#>  0.2966  0.1216 -0.1300 -0.4953 -0.1440  0.1961 -0.8990 -0.2579 -0.7179 -1.5930
#> -2.1795  0.7717 -0.8785 -1.3186  0.0820 -0.8416 -1.7891  0.4833 -0.3553 -0.6409
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]