Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4369  1.4479  1.2163  0.7128  0.1984 -1.3968  0.0101 -1.1880  1.1767  0.1501
#>  0.0450 -0.4556 -0.0041 -0.2410 -0.5117  0.9474 -0.4467  0.8619  1.2349 -1.0321
#>  0.6185 -0.4676  1.1912 -0.1685 -0.7784 -0.6130 -0.8196 -0.9510  0.3117 -1.2659
#> -0.7117  1.0596 -0.3660  0.2494  0.6511 -2.4952 -0.1831  1.1598 -0.3300  0.3915
#>  0.1452 -1.6671 -0.4999 -1.1850  1.2226 -2.0253  0.6965 -0.3724  1.5970  0.8754
#> -2.1545  1.0044 -0.3497  0.1513 -0.2648 -0.3324  0.6583  0.6799 -0.2589  0.2385
#>  1.0101  0.1340 -0.3845  0.8355  0.1922 -0.4527  1.0310  0.2782 -0.0548  0.1779
#>  0.9364 -0.7203  0.8713 -1.1854 -1.7742 -0.7718 -2.2242 -0.7836 -1.8435  2.9336
#>  0.2132 -0.2192  0.8923  0.0148  1.5399 -0.4997 -0.8301 -0.1982 -1.5328  0.1410
#>  0.6326  0.7660  0.2722  0.2336 -0.5976  0.2919  0.5977  0.7971  0.6079 -1.5343
#> -0.4268 -0.9557 -1.1636  1.2456 -1.7095  0.7086  0.8976  1.2038  0.4582  0.0939
#> -0.8654  0.2148  2.0142  0.7879 -0.3425  0.2691 -1.4152  0.6720 -0.9203 -0.7564
#>  0.3696  0.0822 -2.9751  0.7503 -0.2206 -1.0088 -1.1751  1.2855 -0.0701 -1.5920
#> -1.0508  1.5228 -1.4445  1.5362 -1.6315 -1.1909 -2.2097 -0.2137 -0.1849  0.4373
#> -0.5672 -0.2784  0.9020 -0.3012 -1.6728  0.6023  1.1433 -0.4553  0.7120 -0.4742
#>  1.8434  0.5519  0.7676  0.5552 -1.3421 -0.1219  0.3862  1.1436  0.3438  0.1686
#>  1.2289 -1.1433 -0.8951 -1.8514 -0.9357  2.6901  1.3808 -0.1856 -0.7126 -0.7507
#>  0.9816 -1.2768 -0.0345  0.8014 -2.0068 -0.7308  1.0810  0.8207  1.6324  0.9637
#> -1.1061 -0.1878  1.6009  0.1943 -0.2567  0.4131  0.3592 -0.6727  1.3927  0.1684
#> -0.2205 -0.5123 -0.9623 -0.9110  0.1151 -0.5626  1.0791  0.4833 -0.0958 -0.8546
#> -0.6246  0.9431  1.1991  1.3119 -0.3798  0.2881  1.6307 -1.1309  1.6634 -2.0277
#>  0.8837 -0.8360  1.2035 -0.3448 -0.3045 -2.5656  1.0060 -1.4286  0.7335 -0.3516
#>  0.5222 -0.7981  1.8123  1.3349  0.3004 -0.8375  0.9612 -0.3823  0.0448  0.4596
#> -0.4932  0.0507 -0.3123 -0.0329 -0.5753 -0.6759  0.6506  2.2109  1.0969 -0.7064
#>  0.1370 -0.5425 -0.8839 -0.3467  0.2703  1.7709 -1.1639 -0.1166 -0.5829  0.1469
#> -0.3034 -0.7972 -0.5823 -0.2207  1.5713  0.1362 -1.6177  0.1755 -0.6557  1.3055
#> -1.2601  0.3245 -1.0274 -1.3344 -0.1746 -0.9839  0.5225 -1.1540 -1.3969 -0.0550
#>  0.1367  0.4149 -1.0163 -0.9953 -0.5984  0.6054  1.5089 -0.8258  0.0170  1.7391
#> -0.2581 -0.2315 -0.5117 -0.2194  0.4018 -0.3343  0.5069  0.9519 -0.1225 -0.3247
#> -0.5947  1.0898 -0.4496 -0.4103  1.1605 -0.0865  0.9780 -0.2360 -0.0272  0.2783
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]