Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.4271 -2.1204  0.3604 -0.1599  0.0723 -0.0975 -0.1816 -1.7212 -0.1113  0.1120
#>  0.8728  0.8998 -0.2373  2.6230 -0.9361 -0.2560  0.8359  0.0377  1.7334 -1.6145
#>  0.5615 -1.1165  1.0699 -0.3964 -1.2142 -1.5095 -0.8043  1.0288 -0.7458  0.7063
#>  1.1959  0.8328 -0.5005 -1.4572  1.8727  1.8424 -0.1512 -0.3767  0.5866  0.2444
#>  1.3674  1.5500  1.8332  0.0898  2.5536  1.3204 -0.5664  0.8745 -0.8148 -1.3543
#> -1.4636 -0.9746  0.0066 -0.4330  1.4136  0.2575  1.1865 -0.4043 -0.8307 -0.9162
#>  0.0725  1.0156 -0.3110 -0.1579  0.3588 -0.6589 -1.1069 -1.3844  1.0469  0.5616
#> -1.5278 -0.7754  0.9292  0.0906 -1.3385  0.8836  2.3477 -0.4257 -0.6977 -2.2460
#>  2.0110  1.0513  0.9986 -0.5066  1.2729 -1.2295 -0.8873 -0.5466 -0.9505 -0.0464
#> -0.7630 -1.2294  0.4460 -0.7484 -0.2827  0.1071  0.3486  0.5446 -0.7242 -1.1895
#> -0.9644  1.7397  0.0152  0.4172 -0.1219 -1.0527 -0.9960  1.3466  0.5177 -0.6349
#> -1.5455 -0.5891  0.0005 -0.2381  2.1843  0.3191  0.9432  0.2799 -0.5838  2.0249
#> -0.7690  0.5588 -1.7234 -0.8063  0.1731 -1.1117 -0.1076  1.4060 -0.8524  0.3830
#> -1.6872 -0.4798 -0.6990 -1.2366 -0.7172 -0.6908 -0.4676 -0.4104  1.4687 -0.0620
#> -0.2515  0.9187  0.3699 -0.4389  1.3180 -1.1210  0.0892 -0.3598  0.0864 -0.0831
#>  0.8293 -1.0205 -0.3101  0.4461  1.4130  0.9040  1.4078  0.3676  0.0999  1.4430
#>  0.3482 -0.0302  0.3531 -1.2912  1.6503  1.1192 -0.4863 -0.1861  0.0525 -1.0483
#>  1.2126 -0.3165 -1.5917  0.7891 -1.4099  0.7014 -0.1863 -0.2764  0.0539  1.3835
#>  0.6407 -0.6337  0.1718  1.6423  0.2475  1.0082 -0.0268  1.2961  2.0230 -2.4586
#>  0.8511  0.6952 -1.5244  0.3011 -0.4838 -0.4637  0.1823  0.9288  1.1959  0.5056
#>  0.5277 -0.0583 -0.3956 -1.0696 -0.7643 -0.1220  0.5791 -0.0656  0.2877  0.5836
#>  1.9265 -0.4653  0.7289 -0.9505 -0.0408 -0.2207  0.3971 -0.4216  0.4422  0.6637
#>  0.7607 -1.0852 -0.2690  0.4415  0.9617 -0.8332  2.2685 -0.8558  1.0076  1.6015
#> -0.5762 -0.0747 -1.7687  0.9836  1.6919 -0.5940  0.0476 -0.8741  1.5617 -1.2544
#>  0.0022 -2.0078 -0.7655  0.5036  0.3111 -0.4038  0.8511 -0.3256  1.0685  1.7481
#> -0.6068  2.0124 -0.5464  0.7772 -2.0814 -0.7703 -0.0421 -0.3864  0.2430 -2.0593
#>  0.9563  0.8896 -0.4153 -2.1576 -0.7425 -1.4221 -1.1362  0.9324  0.6182  0.0880
#>  0.7266 -2.3707  0.3310  0.4666 -0.9695  0.5556 -0.1723  0.9721  0.3857  0.9834
#>  0.2573  0.8515 -1.5914 -0.5851 -1.4779  0.1404  0.3596 -1.2972  0.3953 -0.9302
#>  2.0583  1.0513 -1.2032 -2.2352 -0.8063  1.1160 -1.8437  1.0183  1.8549  0.7461
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]