Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.3394  0.3097 -0.1804  0.0670 -1.5950  0.5950  0.2830 -1.3065 -0.8980  0.3972
#>  1.3590  0.2290 -0.3840 -0.1340 -1.4259  1.4982 -0.0885 -0.2985  1.8330  1.1304
#>  0.0952 -2.2673 -0.3495 -0.6839  1.7534 -0.1196  0.6157  0.7529 -0.0104 -1.5847
#> -1.5630  1.0511  0.0842  0.4086 -1.7138  0.0921  0.3555 -1.1014  1.4916  0.3862
#>  0.0697 -0.0248  0.3559  0.9927  2.2889  0.0530  0.2585 -0.4012  1.4851  0.3704
#> -0.4318 -0.7271  0.8871  1.7967 -0.2001 -0.4062 -0.0208  0.7050  0.5267 -0.3940
#> -0.8876 -0.4333  0.1695  0.4871 -0.1823  0.5949 -2.2297  1.4585 -0.8875 -0.1832
#>  0.4088  0.7600  0.9858  0.8211  0.7513  1.2334 -1.8025  0.1960 -0.5540 -0.0310
#>  1.7034 -0.4524 -0.6118 -0.7791  0.0411  1.8995  0.5646  0.0936 -1.2748  1.0653
#>  0.4729  2.1149 -1.0492  0.4243 -0.0716 -0.5543  1.0733 -0.8598  0.1191  0.5896
#> -0.1915 -0.2569  0.7634  1.0066 -0.6915 -0.1120 -1.6215 -0.2860 -1.0321  1.4751
#>  0.5549 -0.3378 -0.0244  1.7734  0.4818 -1.1309 -0.2155 -0.4048 -1.5382  1.0838
#> -0.0638  0.7591  2.2201  0.2592 -1.3368  1.4674 -1.7058  0.0658 -1.4696  1.3585
#>  2.2639  0.5634 -1.7445  0.1515 -0.9413  0.2553  1.5374 -1.3074  0.9022  0.6236
#>  1.3862 -0.4180 -0.2026  1.1027  1.0282 -0.0105  1.2813  0.1087  1.1031 -0.3561
#> -2.0269  0.2048 -0.5864  0.1225  1.1500 -1.6121 -0.2080 -0.4614 -0.7473  0.6660
#>  0.3027 -1.3993 -0.8547 -0.3431  0.1633 -0.1057  1.0294  1.8267  0.2530 -1.1867
#> -1.5470  0.8179 -0.5931  1.5877  0.2737  1.4530 -0.2402  1.0872 -0.6893 -0.7021
#>  1.1520 -0.4961 -0.6934 -0.0467  1.4208  0.7247 -0.4804 -1.2589  2.9048  0.2757
#> -0.3372 -1.9818  0.8122  0.3576 -0.4521 -1.5833  0.1579 -0.1622 -0.5417  0.7572
#>  1.9401  0.1291 -1.1637 -1.0875 -1.6650 -0.9987 -1.0661 -0.7760  1.2884  0.7663
#>  3.2760  1.2636 -0.0634  0.0513  1.5046  0.3583  1.3650 -0.0203 -0.5947 -0.5188
#>  1.0736 -0.2802  0.2514  0.1537 -0.8476 -0.1578  0.9977  0.8218  1.2324  0.9367
#> -0.9570  0.1194 -0.1356  0.2356  1.3686 -0.3429  0.7028  0.3218 -0.0880  0.6925
#>  0.7255  1.4201  0.4109  0.7132  0.5118  0.5713  0.7804  2.2257 -0.7019 -0.7559
#>  0.8092 -0.0419  0.4689 -0.5735  1.5053  0.4901 -0.6514  0.8828  0.2696 -0.2672
#> -0.2064  1.1405 -1.7858  0.5821 -0.3284 -0.7882  0.0687  2.2633 -1.2132  2.0587
#>  1.4358  1.6999 -0.9831 -0.5383 -0.0919 -0.1892  0.4350  1.1301 -1.5842 -0.7344
#> -1.7295  0.5294 -1.1211  0.4271  1.1440  0.6035  0.3045  0.9330  0.7932  1.0728
#> -0.4689 -0.9155 -0.3915 -0.1148  0.7532 -0.1693  0.4810  1.3306  0.0721 -0.3848
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]