Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6335  1.4964  0.0838 -1.4758 -0.6206  1.2127  0.0809  0.5266  0.3654 -2.3870
#> -0.5072  2.4363  0.8480  0.4660 -0.0363 -0.6537 -1.4130  0.3269  1.2448  0.3953
#>  1.0093 -1.1748 -1.2521  0.7478 -0.4599  3.2725 -0.0676  1.4418 -1.4051  0.5927
#>  0.8292 -0.1128 -0.2907  0.1832 -1.0580 -0.9330  0.7261 -0.4147 -0.3294  0.5610
#> -0.8344  0.4994  1.3688 -1.0642 -0.4604 -0.7751  2.0435  0.7067 -0.8071 -0.5346
#>  2.6915 -0.6317  0.5130  0.0935 -0.8001 -2.1734 -0.0706 -0.2543 -0.5887 -0.4924
#>  1.3421  0.3398  0.5641  0.6104 -0.5412  0.3190  0.2282  1.4317 -0.0452 -1.0322
#> -1.9099  0.6503 -0.4441  1.6212 -0.3325  0.1500  0.0278  0.2748  0.4191  0.4686
#>  1.3684 -0.2489  0.1099  1.1352  0.1074  2.1294  0.6145  0.0252  0.2136  0.1725
#>  1.0550 -0.7106  1.1995 -1.5316 -0.9075  1.0398 -0.4961 -0.1672 -2.3789  0.0448
#>  0.6002  0.4246  0.4242  0.6907 -2.1013  0.9741 -0.5973  1.6912 -0.0814  1.1917
#>  0.4284 -0.3836  1.1857 -2.1245 -1.0132  0.2031  1.2718  1.5275  0.2279  1.1368
#>  0.9979  1.4107  1.2616  0.2929  1.1677 -0.8092 -0.2090  0.6252  0.5586 -0.7629
#>  0.2384 -0.8212 -1.2086 -1.2816  1.0289  0.5771 -0.6673  0.8488 -0.5119 -0.9173
#> -0.1485  0.6492 -0.4106  1.0071 -1.2118 -1.8007  2.1212 -0.9911  0.2572 -0.5909
#> -0.3752  0.1664 -0.2101 -1.0854  0.0345 -1.0461  0.4729  1.5185  0.3114  0.6454
#>  1.0288 -1.5626 -0.1512  0.7510  1.7636 -0.8373 -0.3396  0.9154 -1.3996  0.4258
#>  0.6516 -0.2470 -1.7834 -2.2518 -0.3406  0.3721 -1.1177 -0.1959  0.2746  1.2121
#> -0.3837  0.3765 -0.1963  1.0454  0.2238  1.1752  0.9485 -0.9574 -1.4956  1.0686
#>  0.1073 -0.3662 -0.6743  1.4303 -0.6284  0.8103 -1.6196  0.0725  1.1277 -0.0771
#>  2.1823 -0.5318 -1.9643 -0.1198 -0.1118 -0.3304  1.5708  0.9880 -0.3301 -0.7364
#>  0.0641  0.2725  0.2104 -0.4031  0.8105  1.1079  0.1271 -0.0117  0.8797  1.5988
#>  2.0411  0.2560 -1.3738 -1.5821  0.8352 -0.5207 -1.3386 -1.1059  0.4360 -1.0741
#>  0.6905  0.3625  0.2524 -0.4806  2.7002 -0.3198 -0.2459  0.7113 -1.0327  0.2582
#> -0.3570  1.3749  0.1188 -0.3669  2.7276 -1.4008  0.7807 -0.3650 -0.9964 -1.1098
#> -0.9405 -0.6333  1.3824  1.2431 -2.9624  0.8894  1.4619 -0.9305  0.8007  0.6350
#>  0.0871  1.8654 -0.2212 -0.1923  0.2125 -0.8996 -0.1796  2.2062 -0.6890  1.7934
#>  0.1562  0.7123 -1.3672  2.5743  0.2762 -1.0474 -1.3774 -2.1926  0.4150 -0.2426
#>  0.4774  1.2506 -0.2415 -0.7565 -0.6987 -0.3376 -1.9436 -0.5411 -1.1001 -0.5976
#>  0.7427 -1.3669  0.9257 -0.9768  0.5798 -0.6231  1.4573 -0.0873 -0.7577 -1.2626
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]