Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0733 -1.1719 -1.5745 -0.5561  1.3760 -1.5180  0.1493 -1.0314  0.2803  0.9843
#>  1.6290  1.3410  0.2029 -0.8529  0.7327  0.0581 -1.2422  0.2576 -0.7780 -0.6317
#>  0.9089  1.4802 -0.7076 -0.2552  1.7023  0.0677  0.7001  2.2465 -0.4219  0.8763
#> -1.6814 -0.2565 -0.2000 -0.4829 -1.1323 -0.8204  0.6609  0.1145  1.2727  1.2194
#> -0.7550 -0.9052  0.6964  0.7730 -2.0690  1.8152 -0.1318 -0.2579  0.0713 -0.6021
#>  0.3247  1.4405 -0.2981  1.2690 -0.2827  0.5189  0.4792 -0.1380  0.7733 -1.7171
#>  1.6959 -0.6374  0.7762 -0.3827  0.0805  0.5746 -1.1045  0.9456 -0.9616  1.1146
#>  0.1767 -1.6882 -0.4497  0.5569 -0.5409  0.0494 -1.4162  0.1755  0.1622  0.7195
#>  0.5361  0.1809 -0.7832 -1.0734 -2.0938 -0.2431  0.7047  0.6624  0.5499 -2.6614
#>  0.7620 -1.2214 -0.3060  0.2037  0.2054 -0.9963  1.3825  0.5428 -1.2476  0.0981
#>  0.7563  0.2857  0.2235  2.4203 -2.4903  1.3751  0.0090 -1.2225  0.5714  0.4974
#> -1.2498  0.0165  0.7612  0.2649  1.8389 -0.1346 -1.8636  0.1051 -0.0213 -0.1188
#> -0.5898 -0.3774 -0.2789 -1.3939 -0.0532  0.1543 -1.4055 -1.7617  0.9682 -1.2061
#> -1.5283 -1.1537 -0.3813 -1.7727  2.2420  0.2111 -0.6035 -2.4946 -1.0517  0.1892
#>  0.9800  0.4038  0.4796  0.0898 -0.0467 -0.3954 -1.5483 -0.2754 -1.5511  1.4711
#> -0.2516  0.5018  0.6670 -1.6323 -2.0728  0.3829 -0.2534 -0.1169 -0.7795  0.8818
#> -1.0827 -0.7728 -0.6653  0.8383 -0.0341  0.3196 -0.0174  1.4591  0.1696  0.0390
#> -2.0348  0.6281  0.0186  0.5211 -0.2079  0.5667  0.5387 -1.5630  1.2430 -0.7876
#> -0.4437 -0.4111  1.4022 -0.2327  1.2436 -0.1534 -1.0179 -0.2157  0.1284  0.2735
#> -1.5371 -0.5923 -0.0526 -0.9331 -1.5350 -0.9645  1.4763  0.2821 -1.9206  1.0821
#>  0.9419  0.0039 -0.8182 -2.4470 -0.1434 -0.2791  0.5768 -1.7221  0.2023  0.9285
#> -1.3979  0.8385 -0.2313  0.2208 -0.0791 -0.2631  0.8581 -1.7501 -0.4496 -0.8718
#> -1.2973  0.0914  0.4053 -0.0635 -0.4747  0.6844  0.9530  0.5596  1.4887 -0.6503
#> -1.5306  0.8402 -0.9067  0.1149  0.0714 -0.5842 -0.8615  1.6097  1.3817  0.4200
#> -2.0742  0.2379 -0.4511 -1.1858 -0.1560  1.4073 -0.1278 -2.7421 -0.9151 -0.9817
#>  1.2254  0.4557  0.5608  0.1709  1.0947  1.4423 -0.5016 -0.0226  0.9567 -0.3366
#>  1.0715  0.9621  0.1120  0.1831 -0.7369 -2.0686  1.2762 -0.6509 -0.3500  0.7963
#> -0.2161 -0.0226  0.5907 -0.8178  2.4902 -1.2963  0.8115 -0.4670  0.8633 -0.5815
#> -0.0123 -0.0958  1.1848 -0.6007  1.4862 -0.0319  1.3433  0.0786  0.0875  1.0761
#>  0.0663 -0.8815  0.1047  0.2250 -1.9819  1.4482  0.9500 -0.3877  1.1047 -1.2721
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]