Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7414 -0.4636  1.4589 -0.1548  1.7606  0.3783 -2.3777  0.2589 -0.0745  0.2741
#>  1.9943 -0.0083 -1.1580  0.1971  0.0778  0.3604 -1.0420 -0.1773  0.6873  1.2588
#> -0.9719 -1.0865  0.0349 -0.0720 -0.3503  0.5060 -0.5263  0.0177  0.3926  1.4846
#>  1.7189 -0.6878 -0.0009  0.5583  0.4102  0.3688 -0.7188 -0.0791 -1.5195 -1.4432
#>  2.3641 -2.3560 -0.3278  0.5467 -0.3626  1.4628  1.3161 -2.1733 -0.2153 -0.9533
#>  0.3605 -0.5436  1.0040 -0.2582 -0.9393 -1.1197  2.0244 -1.0091 -1.6106 -1.2685
#> -0.6624  0.8330  0.9180 -0.2972 -0.0285 -0.5599  0.4243 -1.0893  1.2079  0.2867
#>  0.0766  0.4253  0.6731  0.2205  0.3621 -0.1889 -0.1080 -2.1519 -0.3228  0.4023
#>  0.2321 -1.7461  0.0891  2.3865 -0.4590 -0.2024 -1.1696 -1.5874 -0.2775  0.4396
#>  0.8252 -2.7657 -0.2100  0.8290 -0.7408  0.8524  2.1456 -2.9032  0.0261  0.3371
#>  0.2102  1.1705 -0.1915 -0.6523 -0.0866 -1.5368 -0.6078  1.5487  1.1029  1.0754
#> -1.2586  0.8253 -0.2107  0.3240 -0.6033  0.0381 -0.8639 -0.3161  0.4422  0.7529
#> -0.1341  2.2022 -0.2806 -0.5528 -0.5378  0.9802 -0.5749 -0.6116 -1.3815  1.5574
#>  0.3330  0.9655 -0.7363 -1.3088  0.2192 -0.9700 -1.5764 -1.3581  0.2870 -0.0781
#>  0.3814  1.1498  0.3606  1.2334 -0.3804  0.7892 -1.0156  1.6152 -0.2009  1.3003
#> -0.6469 -1.7890 -1.6750  1.4492  1.0517  0.2728  0.7503  0.8211 -0.9347  1.1249
#> -1.4812  0.7251 -1.4291 -1.2257  1.4495 -0.2418 -0.2053 -0.1924  0.3835  0.3481
#> -0.8038 -0.3984 -0.9355  0.6486  0.6459  0.9656 -0.5006 -0.3854 -2.2837 -1.3003
#>  0.2750  1.0535  0.6026  0.5010  0.5224  0.0474 -1.1501 -0.3819  0.3111 -0.0865
#> -0.5583 -0.5055 -0.5019  1.8396  0.7102  0.0061 -0.9661 -0.0010  0.2890  1.3195
#> -0.0235 -1.1704 -0.2661 -0.5475  0.2998  0.3381  1.5654  0.9671  0.0596  2.9138
#> -0.0925 -0.6505  0.6358 -1.0614  0.1473  0.2543 -0.1142 -0.5191 -2.0240  0.2100
#> -1.9498  0.3785  0.4284 -0.0569 -0.2507 -0.3319  0.3615  0.7343  0.1684  1.1606
#>  0.8497 -0.5510  0.8104  2.9493 -0.0479  0.1732  1.1604  0.0696 -0.4528  0.6459
#>  0.5108  0.6559 -0.2730  0.2888  0.6567 -0.5262  0.2378 -0.7825  0.5435 -0.6523
#> -1.1216  2.0250  1.6941 -0.0937 -0.3394  2.0780 -1.3054  0.4328 -1.3620  0.5752
#>  1.0600 -1.1248 -1.4961 -0.1057  1.5896 -0.9303 -0.5641  1.3109  0.5335  0.5694
#> -1.3202  1.9764  1.5330 -0.3865 -0.6127  2.8688 -0.1849  0.2498  0.2727 -0.7745
#> -1.4928  1.1321 -0.5859  0.7529  0.5430  0.1794  0.6343  0.0535 -0.1961 -0.6448
#> -0.0838 -0.1306  0.3429 -1.0946 -0.2875  0.1217  0.0317  0.9322  2.1278  0.5938
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]