Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0705 -0.4519 -0.2268  1.4005  0.1043 -0.2564 -1.0834  0.8869  0.3397 -0.6338
#> -0.5176  0.9251 -0.4373 -0.7891  0.3649 -1.0339 -1.3472  0.6937 -0.7874  0.9922
#>  1.3783  0.1938  0.0405 -0.4538  0.2693 -1.5077  0.3244 -0.2421  0.9046 -1.1681
#>  0.7967  0.1852 -2.0542  0.7434 -0.7137  1.0743  1.4078 -0.8986  0.0591 -0.4927
#>  0.3787 -1.5681 -0.5216 -1.4917 -0.9988  0.1980  0.5476 -1.3366 -1.1380  1.8701
#>  0.2825 -0.9784  0.5799  1.6863  0.7050  0.0674 -0.2815 -1.1732  1.1320 -0.1695
#> -0.2457  0.0901 -0.0476  1.0230  1.0189 -0.2075  0.9166 -1.3160  1.3645 -0.0448
#> -0.2042  0.6050  0.1133 -0.0450  0.0011  0.4472  0.4962  0.0769 -0.3430 -1.9411
#> -0.9237  1.3455  0.2376 -0.8501 -0.5261  0.8713 -1.4416 -0.5493  0.1909 -0.4490
#> -0.2517  0.3250 -0.9983 -2.2586 -1.3742 -0.4778 -2.0711  0.5664  0.0223  1.1653
#>  1.2537 -1.9554 -0.4129 -0.6900 -2.2394 -0.3216 -0.6191  0.2452  0.1787 -0.2626
#>  0.7943 -0.5497  0.4382  1.2424 -1.0839 -0.6827  1.2189 -0.4982  1.6120  1.6883
#>  1.2391  1.8074  1.2741 -2.8368  0.7268 -0.9855  0.4233 -1.0154  0.9302  0.6127
#>  0.5540  2.1198 -0.0819  1.0926  0.1972  0.7682 -1.3632 -0.0204  0.1875  0.0547
#>  0.0011  1.4671  0.2932  0.4870  0.5507  0.3313 -1.0224  0.5472  0.7251  0.2538
#>  0.8981 -0.3559  0.6572 -1.1679  0.8774  1.1777 -0.2516 -0.8758  0.3892  0.7249
#> -1.6689 -0.5736  0.4394 -1.2014 -0.7546  1.7978 -0.4983 -1.1159  0.4740 -0.2861
#>  1.7321 -0.7244 -0.9051  1.4128  0.4684  0.5576 -1.4037 -1.5375 -0.3149  0.0358
#> -0.1160  0.1352 -0.0361 -1.3672  0.6059  0.0732  1.0242 -0.2786 -0.0062 -1.7050
#>  2.4150 -0.4379 -1.5817 -1.0022 -0.2953  1.5481 -0.7706  0.9163  0.4715 -0.9739
#>  1.2348 -0.9801  0.1748 -0.9443  0.1111 -0.8692 -1.5706 -0.8057 -0.6050 -0.1486
#> -0.3675 -0.3961 -0.2095 -0.6929  0.6116  0.2996 -0.2592  0.7818  0.7146  1.4617
#> -0.2472  1.6802  1.5555  0.0798  1.6876  0.0528 -0.0241  0.3806 -0.0272  0.2863
#>  0.3396 -0.8870 -1.8541  0.5599 -1.2129 -0.2844  0.3850 -0.8815  0.0599  1.1740
#> -0.8075  0.6235 -0.5100 -0.3790  0.1756 -0.7309  0.2164 -1.3444  0.9279  0.3616
#> -0.9100  1.7366  1.9241  0.8810 -0.7127  0.4866 -1.0625 -2.5763  0.8117 -0.3246
#> -0.0509 -0.0604 -0.7506  0.3947 -1.4813 -0.8628 -2.4517  0.0637  1.0239 -1.1579
#> -0.8677 -0.3692  0.5513  0.9966  1.1211 -0.3061  0.0394  0.1792  0.8696  0.7867
#> -1.8457 -0.1101 -2.2461 -0.1517 -0.3821  1.1997 -0.0254  0.3268  1.7199  0.1905
#>  0.2965 -1.5785  0.9402 -0.1982 -0.2475 -1.5056  0.5889  0.7824  1.5703  0.2785
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]