Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.0980  0.2430 -1.7894  1.5223 -1.7997 -0.8695 -1.1285 -0.8120 -2.1036 -1.1122
#>  0.2741 -1.0266 -0.1776  0.5556  0.7384  0.5334 -0.0080 -1.3279 -0.2820  0.0150
#> -0.9513 -0.8011 -0.4686 -0.2677 -0.4389  0.1731  0.1128 -0.1582 -0.2677 -0.9230
#>  1.1695  0.1567  0.6430 -0.0722  0.2322 -2.1179 -0.2377 -0.1803 -1.2841 -0.0093
#>  1.3269 -1.0978  1.1728  0.7998  0.6046 -0.5591 -0.4029 -1.6213  1.2932 -0.9690
#> -1.0258  0.5104 -0.1019  1.1617 -0.2230  0.9862  0.9121 -0.4959 -0.9572  0.1964
#> -0.2127  0.0148 -0.8972 -0.7077  0.7602 -1.0714 -0.0864  1.1334 -0.7961  0.3083
#>  0.1946 -0.3735 -0.2326  0.9253  0.1429 -1.5561 -0.0678  0.1433 -0.8669 -0.4351
#>  1.4883 -0.2313  0.1678  0.7233 -0.2444  0.2761 -1.5166 -0.9023 -1.7422  1.0795
#> -0.3949 -1.2572  0.3676 -0.6579  0.9493 -0.5699  1.3134 -0.7564 -0.1440  1.1959
#>  0.8646 -0.4744  1.0465 -0.7408  0.0772  2.0174  1.5352 -0.0463 -2.3375  0.7602
#> -1.0292  0.6782 -1.0058  0.1978 -0.8358  0.0133  0.1528  1.4461 -1.4043  0.0432
#> -0.7257 -1.5677 -0.3472 -0.0743  0.7191  0.0250 -0.6992 -1.0151 -0.1155  0.3449
#>  1.9556  0.9283  0.7800  0.1531  0.1736 -1.7214 -0.1222  0.7597  0.7998 -1.2736
#> -0.8738 -1.1610  0.2756  0.8091 -2.0336 -0.4444  0.4014  2.2415  0.6717 -1.4919
#>  1.2133  1.3821  2.0035  1.4062  0.8131  0.9068  0.7776 -0.2689 -1.2875  0.0484
#>  1.2694 -0.5093 -0.5706  2.0767  1.0318 -1.3611  2.6467 -1.7817  2.2306  0.7618
#>  0.2594  0.5399  1.8382  0.3715  1.4232  0.1887 -0.8407 -0.2826  0.4203  0.7725
#> -0.0560 -0.1090  0.1347 -0.3772 -0.9707 -1.3468  0.0575 -0.3588 -0.7435 -1.2772
#> -0.1961  0.0458 -0.2394 -0.4662  0.0692 -0.4108 -0.9505 -1.8993 -0.5130  0.7388
#>  0.1529 -0.1703 -0.7216  0.5216 -0.8867 -1.9867  0.3137  1.7253 -1.1598 -1.5119
#>  0.1382 -0.9685 -1.1236 -0.6154 -0.8823 -1.3735 -1.2388  0.2532 -0.1473  0.1967
#> -0.0167 -2.2623 -1.7036  0.5940  0.3905 -0.2693 -0.5447 -0.3880 -0.2692 -0.4320
#>  0.2431 -0.8040 -1.2589  0.9396  0.4113 -0.3225  0.1404  0.2039 -2.0292 -2.0168
#>  1.7746 -0.9709  0.8714 -1.2067  0.8733 -1.8608  0.7271 -1.2227  1.6562 -0.1019
#> -0.2790 -1.1666  0.8194 -0.8420  1.2615 -0.1123  0.6062 -0.4867  0.1886  1.1218
#> -0.2108 -0.4668  0.6019 -2.9551  1.8133 -1.1412 -0.9481 -1.0619  1.1267  0.5981
#>  1.2103  0.8679 -0.3753  0.1938 -0.0354  1.5706 -2.2738  0.2751 -0.5409  0.4165
#>  0.3510  1.5886 -0.0529  0.8642 -0.2329  0.0302 -0.9945 -1.1662 -1.4963 -0.7607
#>  0.0101 -1.1300  0.3824 -0.9959  1.3015 -0.8923 -0.2832  0.7085 -0.9830 -0.3282
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]