Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0148 -1.5245 -0.9772 -1.4405  0.7108  0.7032 -1.9035 -1.0581  0.3833 -1.4236
#>  0.8820 -1.8555  0.4716 -1.2107 -1.4123 -0.0234  0.1141 -0.1081 -0.3140 -1.1781
#>  0.8270 -1.4031 -1.7363  0.3312 -0.0387 -0.2023  0.7634 -0.0041 -0.7231 -0.5782
#> -0.7197 -0.7068 -0.4122 -0.3529 -1.0571 -0.0236 -0.9859 -1.7557 -0.8312  1.0551
#>  0.6672 -2.5029  0.0878  1.1782  1.1054  1.1809  1.3533 -2.3741 -1.9601  1.3502
#>  0.3081 -1.4134  2.2348  0.1056 -0.5324  0.6270 -1.0937  0.7005 -2.1779 -1.7547
#>  0.4827  0.2632 -1.2270 -0.1591  0.1170 -0.1513 -0.1581 -0.4925 -0.2791 -0.4849
#> -1.4595 -0.1817 -3.2769  0.0397 -0.1107  0.2290  0.1939  0.6403 -0.6313 -0.9118
#> -0.4043  0.6134  0.0397  0.5266  0.0082  0.2495 -0.4159  1.5716 -0.4589  0.3856
#> -0.2030 -0.9708  0.2421 -1.9933 -0.0977 -1.9690  1.1027  1.0931 -0.4236  0.3603
#>  0.4785  1.7938 -1.2018 -0.8216 -1.1747  0.1497  0.7935  0.3276  2.4088  2.1383
#> -0.0483  0.9016 -0.0345  0.0230 -1.3112  0.8159 -1.6895  0.0618 -0.2764  0.7672
#> -0.2062  0.9615  0.7508 -1.2316 -0.8073 -0.4681 -1.7652  0.1735 -1.0109  1.2113
#>  0.9530  0.8128  0.7152  0.7007  0.8974  0.7841  0.6819 -1.2762 -0.5009 -0.5403
#> -0.9151  0.1054 -1.0753 -0.1976  1.3599  1.0225 -0.6925 -0.1314 -0.6412 -0.0018
#> -0.5531  1.2693 -0.1625 -0.5296 -1.0258  1.3637  0.4432 -0.4573  1.5607 -0.0561
#>  0.2206  1.6025 -1.1916  0.5576  0.4835 -0.1972  0.9392  0.3389 -1.0885  0.9408
#>  0.5982  1.5313  0.2971 -1.8858  0.1569 -0.4810 -1.8325 -0.0816 -0.2190  0.6041
#>  0.2240  0.7121 -0.8101  1.5472  0.0110  0.0080  0.9919 -1.6764 -0.1495  0.4819
#> -0.5939  1.1267  0.4555  0.6344 -1.8257  0.6129  0.1164  1.9882  1.5376  0.6375
#>  0.3599  1.5712 -0.5940 -0.4970  1.8630  0.0418  0.5060 -0.5717  1.1914 -1.2764
#> -1.0685 -0.4097 -0.2159 -0.2759  0.4539  0.3515  0.3326  0.7335  0.9874 -2.0447
#>  0.0321  0.5755  0.5389  0.4719  0.1470 -0.2720  0.1309  0.5684 -0.0464 -0.6439
#>  0.4403 -2.2948  1.9860  1.6886 -1.4123 -1.2254  0.4669 -0.8954 -0.0142 -0.4962
#> -0.9265 -0.7969  0.2099  0.0523  0.1670 -1.6605 -0.5921 -0.5168  0.6075 -0.2397
#>  0.3853  0.2905  0.9959 -0.2344 -0.6879 -0.1011 -0.1550  1.0850  0.0183  1.0846
#>  0.5191  0.6480  0.3680 -1.9366  1.9661 -0.0508 -1.5726  0.3169  0.3524  0.3843
#>  0.3969 -0.0865  1.5868  0.3371 -1.0024 -0.0937 -1.3659 -0.5371 -0.2522 -0.1032
#>  0.7435 -0.4272  0.7355  1.5812 -0.0371  1.0452  1.5817  1.1683  1.1263 -0.6320
#> -0.6267  0.6688 -0.2140 -0.8917 -1.5734 -0.2692  0.8802  0.4798 -2.0107  0.0684
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]