Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4084 -0.1212  2.7124  0.1693  3.6958  1.1886 -0.8382 -0.9621 -0.4477  0.8767
#>  0.1403  1.0154 -0.1254 -0.1114 -2.2414 -1.6004  0.4117 -1.6845 -0.1337  1.8380
#> -1.5904 -1.1772 -0.8355  1.0961 -1.0823 -0.8774  1.1967 -1.3209 -0.5172 -1.4753
#>  0.1113 -2.0043 -1.1247 -0.6566  0.6687  0.4796  0.3135  0.7370 -0.1250  0.1982
#> -0.7407 -0.1965  0.1516 -0.7867 -0.1543  0.1018 -0.6619  1.0576 -0.0970  1.5939
#> -0.0020  0.8926  0.3103  0.4870 -1.1349  0.0131 -1.1885  1.1342 -0.6589 -0.2797
#> -0.4375 -1.0216  0.1216  0.7083 -0.0817  1.7898 -0.1478 -2.0065  1.1804 -0.9489
#> -0.6821  0.0704  1.0846  0.0987  0.5230 -1.4467 -0.6037  0.9887  0.9435 -1.4225
#> -1.0123  0.9165  0.4327 -0.6730 -1.8465  0.8397  1.0876 -0.1088 -0.5667 -1.7577
#>  0.2606  0.0004 -0.7573 -0.6886 -1.6477  0.4056 -0.1589  1.9904  1.6396 -1.1268
#> -0.8192 -0.7875  0.5664 -0.4422 -0.4877  0.5888  0.2796  1.6005  0.4377 -0.5096
#>  0.3569  0.2458  1.7836  0.4752  1.0538  0.0033 -1.9568 -0.1270 -1.0016 -1.6842
#>  1.2057  1.4288  1.4379  0.4199 -0.8936 -0.3677  0.7518 -0.9957 -0.6988  0.8979
#> -0.1733  1.6845 -0.2837 -0.4023  1.1160 -1.0011 -2.3681  0.6098 -0.0351  0.0385
#>  0.4635 -2.7257 -0.7818 -0.1966  1.4896  0.2997  0.0643  0.3961 -0.1885  0.9004
#>  0.1915  0.2931  1.2371 -1.2550 -1.6768  1.7617 -0.1608 -0.3937 -0.0653 -0.6860
#>  0.3299 -0.6578 -0.0618 -0.1252 -0.5262 -0.0858  2.3279 -0.2788 -1.8656  0.7693
#>  0.8850 -1.1434  1.3971  0.6718 -0.6583  2.8086 -1.3406  0.0925  1.0070 -0.9897
#>  1.4277 -0.3970 -0.6013  0.5945 -1.5324  0.3263 -1.2369  0.2214  0.0964 -2.0216
#>  0.5874 -1.3866  1.7158  0.7920 -0.0679 -0.9943 -0.0478 -1.5918  0.4689  0.6566
#>  1.3125  0.4719  0.4727  0.4299 -0.0927 -0.0108 -1.4023 -1.0825  3.8578 -1.6020
#> -1.3188 -1.5888 -1.0480 -0.2839  0.1358  0.9663 -2.1147 -0.1479 -0.0235  1.0594
#> -0.3188 -1.1802  0.7132 -0.7913  1.2189 -1.3406  0.1699  0.8156 -0.8366 -0.4878
#> -1.2673 -0.7777  0.4892 -0.7224 -1.6801 -0.2391 -2.0736 -1.5426 -0.7324  0.2872
#>  0.5475  0.0464  1.8050  0.1335  0.2269 -0.1408  0.8290 -1.3809  1.0640 -1.7151
#>  0.3421 -0.0246 -0.0752  0.1491  1.3955 -1.5242 -0.0526 -0.1644  0.1489  0.7651
#> -1.4731 -0.9918 -0.4906 -0.0752 -1.4762  0.4665  0.2137 -0.8532  1.2971  0.2083
#> -2.6333 -1.2838  0.3028 -0.7632  0.0772 -1.8677 -0.1065  0.1710  1.8540 -0.3440
#> -0.4962 -1.2746 -1.3129 -0.1710 -0.6556  1.6778 -1.2054 -0.4975  1.3711  0.7420
#>  0.4996  0.4755 -1.2555  1.6436  0.6253  0.5122  1.1244 -1.7371 -0.1628  0.9198
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]