Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7520 -0.4193 -0.1782 -1.8240  0.8561 -1.1951 -0.3777  0.1257  1.0595  0.3775
#> -1.5217 -0.4589 -0.7851 -0.6231 -0.1240  0.5504 -0.1925 -0.9421 -0.6222  0.0385
#>  0.4639  0.4810  0.8586  0.9474 -0.2485 -0.2471  0.2176 -0.3624 -0.8833 -0.9420
#>  0.1320  1.1617  0.7669  1.3205  0.7380 -1.3163 -0.2770 -2.2088 -1.3271  2.8742
#>  3.0648  0.5763 -0.6969 -0.5271 -0.7885 -1.0557  0.5692  0.3799  0.0707 -0.1188
#>  0.2827 -0.7097  2.0983 -1.2569 -0.6174 -0.1389  0.6023 -1.0403  0.8935 -0.0313
#>  1.3078 -0.7659  0.9097  1.9247 -0.1016  0.1026  2.8982 -0.4790  1.2817  0.0679
#> -0.7829  0.3543 -0.0365  0.2326  0.5193 -0.5367  1.4356  1.1979  0.0474 -0.3203
#>  0.6832  0.1880 -0.0293  0.2135  0.9124  1.2148 -0.0839 -0.0911  0.8561  1.5766
#>  0.1300 -0.4375  0.7593  0.2907 -1.6375  0.5250  1.1539 -0.1823 -1.4928 -0.1245
#>  0.0119  0.0968 -0.7271  0.1652  0.5031  1.7120  0.3239  0.3625 -0.6082 -1.9544
#> -0.8879  1.0459  0.5216 -0.7228 -1.7492  1.2092 -0.9754 -0.3497  1.6700 -1.0355
#> -0.6383  0.0584 -0.4501  0.9501 -0.2354 -1.1267  0.1696 -0.0338  1.0647 -0.4061
#>  0.2895 -1.3592 -1.7418 -0.4136 -0.8447 -1.4424 -1.0292  0.5377 -0.9607  0.6545
#> -0.7593 -0.8973 -0.7461 -0.3115 -0.2643  0.9669 -1.0587  0.2097  3.1550 -0.7921
#>  1.3001  1.0038  1.7825  1.9584 -0.4046 -0.0088  0.7803  0.7066  0.3766  0.7375
#>  0.9968 -0.9629 -0.3978  1.4963 -0.3023  0.4868 -1.0364  0.1008 -0.4280  0.1911
#>  0.8714  0.8116  1.1580  0.7815  0.3559 -1.6324  1.5840 -0.9931 -0.9454  1.5472
#>  0.3943  1.0509 -1.2980 -0.0847 -1.1689  0.9580  2.2046  0.0561 -0.2261  0.5668
#> -0.0126  0.2415 -1.7307  0.5399 -0.3315  1.2929  0.8476 -0.9363 -0.5453  0.9857
#> -1.7638  1.3403  0.1990  0.0374  0.5370  1.1418 -0.5093  0.5857 -1.9579  1.3245
#>  1.6980 -0.5095  1.9442 -0.3964 -1.6028 -0.6595 -1.1404  0.7567  0.1356  0.6751
#> -0.2669  0.8029  0.5188 -2.0757  0.5592 -0.9012 -0.0410 -0.5436 -0.3286  1.0108
#> -1.0064  1.1697 -0.1484 -1.2416  0.0467 -2.0526 -1.4425 -0.4824  0.2425 -1.1892
#>  1.2380  1.0431  0.9869 -0.2656  0.4132 -1.1294 -0.2132  0.4311 -0.1227 -1.1822
#>  0.8921  0.5311  1.1501 -1.3479  0.3721 -0.9555 -1.3592 -0.7169 -0.9354 -0.0720
#> -1.1282 -1.0236 -1.0599  0.3608  0.4686  0.8575 -0.2602 -1.5971  0.2525 -0.3519
#> -0.5720  1.7200 -0.7075  0.3319 -0.7363  0.7116  0.7299 -0.4345 -1.3705  1.1908
#>  0.5295 -1.3471 -1.0804  0.0618  0.0446 -2.0512  1.2669 -1.4768  0.2904  0.6852
#> -1.0451 -0.2723 -0.7746 -0.3857  0.1298  0.5777 -0.9142 -0.0153  0.9305  1.1314
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]