Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3352  0.5478  0.4741 -1.9865  1.0685 -0.2178 -1.7603  0.6840 -0.0739 -0.3379
#> -0.2385  0.3848  0.0601 -1.7647 -1.4823 -0.6900  1.9908  2.1125 -0.4318  0.2403
#> -0.1586 -1.1531 -1.0664 -1.5241  1.1604 -0.9614  0.6960  0.2782 -0.6830  0.6806
#>  1.4843  0.0197 -0.2591 -0.8713  1.4904 -0.6804 -2.1918 -0.0152 -1.6483  0.0238
#>  0.7106 -0.6051 -0.9138  1.0244 -0.6067  2.3463 -1.3893 -1.2842 -0.5897  0.9812
#>  0.3405 -0.3068 -0.2217 -0.5898 -0.5273  0.3478 -1.0288  0.1144 -0.7325 -0.9701
#> -1.4014 -2.5581 -0.6558  0.2729 -1.3670  0.9132 -0.0675 -1.9597 -2.2316  0.5992
#> -0.2901  0.8424  0.1279 -2.2949  0.4745 -0.6016  0.5578  1.7013 -1.6577 -1.4232
#> -1.7859 -1.6803  0.2888  0.1551 -0.1195 -0.4919 -1.2458 -0.5847 -0.3431  1.2352
#>  0.3176  1.2297  0.4693  2.3323  0.6664 -0.7796  1.6270 -1.3002 -1.0236  0.2606
#>  1.5578 -1.0266  0.4800  1.2395 -0.5578 -0.3267 -0.4377  0.6999  0.0721 -0.0291
#> -0.6144  0.6265  1.3361 -0.8750 -1.7885 -0.0814  2.2112 -1.1953 -0.7857  0.7618
#> -0.6247  0.5456 -0.4343 -0.5672 -0.9794 -1.2951 -0.6163 -0.8026  0.4628  1.0756
#> -1.0751 -0.2930  0.8432 -1.2367 -0.2130 -0.5862 -0.0565 -0.7393  0.9032 -0.2827
#>  3.1949  0.1040 -0.5129  1.0883  0.8839 -0.0160 -0.4720  1.7593  0.5633  0.9787
#> -0.2074  0.4128  0.4870 -1.4686  0.1465  0.5381 -1.0935 -2.0677  1.4396  0.6570
#> -0.5230 -0.0258  0.4383 -0.7973 -1.1845  0.3068 -0.1733 -2.1504  0.4035 -1.0959
#>  0.1049 -0.2937 -0.7946 -0.1403  0.9081  1.2726 -0.2747  0.7454 -0.1289 -0.3484
#>  1.2912 -0.4438  0.1270  0.3272 -0.3120 -1.6438 -1.2955  0.1442 -0.2320  0.2440
#> -1.2384 -0.9198 -1.6792  2.3409 -0.0186 -0.3143 -0.8894  0.3930 -0.7190 -0.3474
#>  0.0499 -0.2490  0.3775 -0.6637  0.7200 -1.4709 -0.0093 -1.6256 -0.4160  0.2480
#>  0.2955  0.0611 -0.5699  0.1158  1.4005 -0.6332  1.6251  0.7038 -1.4722 -1.0648
#>  0.7070  0.2986  0.8401  0.2248 -0.8413 -1.7619  1.6190  0.0406  2.6638  0.3083
#> -0.1524 -1.2484 -0.2565 -0.5861  1.4714 -0.4865  0.0691 -0.0049  0.9055 -1.1690
#>  0.8922 -0.5893  0.0843  0.4373 -0.0508 -0.4998  0.3368  1.5764  1.5885 -1.0476
#> -0.1837  0.6249 -0.4665  1.0642  0.4135 -0.4048 -0.5314 -0.3307  0.4218  0.6492
#> -0.0342  0.8692  2.0280  1.2034  0.4288 -0.7715 -0.1886 -0.5558  1.0356 -0.9773
#>  1.0615 -0.5894 -1.4517  0.0306  1.2323 -1.0653 -0.2681  3.5767  0.5139 -0.8204
#>  1.4505 -0.5241  0.6081 -1.0785 -1.6928 -0.3005 -0.9643 -0.1168 -0.1618 -0.0774
#> -0.7675  0.1289  0.2049  0.9717  0.3830 -0.3374  0.7269  0.8157 -0.6954 -0.2099
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]