Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.4221  0.6514  0.0596  0.1337 -0.8812 -1.3789  0.2025  0.6848 -0.9528 -0.8739
#>  0.9088  0.5076  0.5636  0.3844  1.6844 -0.5082 -0.5187 -0.8605  0.1265  0.2528
#>  0.4187 -0.0521  0.4961 -0.5604 -0.6025  0.4540  0.0687 -1.1916 -1.0008  0.0931
#> -1.7371  0.4588 -1.1648  0.6741  0.6866 -0.1046  0.5670 -0.4112 -1.1431 -1.0490
#> -0.3840  1.1338  0.4069  0.2322  1.0640  1.0151  0.1627 -1.0463  0.8854  0.2652
#>  0.0622 -0.0529 -0.7380 -0.1129 -0.9308  1.2292  0.0218 -1.2038  1.9014  0.4478
#> -1.3600  0.9587  1.2648  0.7005  0.4566  0.4998 -0.8101  0.8959 -2.2071  1.6045
#>  1.6816  1.1915  0.0722 -2.7645 -2.1086  0.4534 -0.0330  0.9339 -1.6330 -1.3024
#>  0.2125  0.7036  2.1696  0.3929  1.7592 -1.2747  0.2316  1.0097  0.0736  0.6654
#>  1.0189  0.1706  0.7293  0.6143 -1.2509 -0.8409  1.4323 -0.1439 -0.3648 -0.7977
#> -2.7724 -0.3910 -0.6828  0.9218 -0.4679  1.9225 -0.3158 -1.0182 -1.0516 -0.7653
#>  0.2118 -1.0642 -0.1623 -1.2303  0.5650 -0.4506  0.7733 -2.3691  0.9311 -0.1392
#> -0.3985  0.4466  0.7517 -0.4733 -0.1011  0.0492 -2.8746 -0.7075  1.3834 -0.6983
#>  0.6822  0.8770  0.8089  0.9925  0.5977  0.4423 -1.2318  0.8322  0.7339 -0.1345
#> -1.2228  0.0880  1.2364  0.0946 -0.6973 -0.3385  0.0585  1.3016  1.2588  1.2267
#> -0.7263 -0.9868 -0.8598 -0.2024 -0.0299 -1.5691  0.1846 -0.0021 -1.0496  1.3158
#>  0.2768 -0.2272  0.1458  0.1312 -0.7820 -0.3379  0.4080 -1.0879 -1.1183 -1.1261
#> -1.4104  0.8520 -0.1400  1.1386  1.5073 -0.1480  0.8641  0.3384 -0.0405  0.3260
#>  0.6734 -0.2394 -0.0877  1.5036 -1.4012  0.7809  1.7407  1.8115  1.1688  1.0694
#>  1.3417  1.3310  1.2179  1.3646 -0.4921  1.1385  0.4709 -1.2955  0.5648  0.2441
#>  0.4108 -0.7641  0.3179  0.5536 -0.8141 -0.4541  0.3794 -2.6398 -0.2648  1.4454
#> -0.6943 -0.2393  1.4976 -1.0068 -0.6697 -1.8566  1.5308  0.6250 -0.1337 -1.0717
#> -0.0375 -0.4873  0.2062 -1.7324 -1.2201  0.1389  0.2183 -0.1263  2.4931  0.2057
#>  0.3975 -0.6054 -0.9336 -0.5228 -0.3129  0.4918  0.1448  1.5221  0.8001 -0.2893
#> -1.2160 -0.1785  0.4394 -3.1642  1.0070  0.5860  0.0946  0.3215  0.1897 -0.4556
#>  1.5978 -1.2491  1.0685  0.9974 -0.6770 -1.2392  0.3837 -0.9758 -0.2232 -0.2535
#>  0.3102  1.6805 -0.2510  0.5617  0.2540 -0.0183 -1.6769  0.3324  0.4224 -0.4118
#> -0.0004  1.0650  1.6985  0.2352 -1.6874 -1.2233 -0.5368  0.3108  0.4624  0.1361
#>  0.9617  1.4017 -0.4987 -0.2719  0.2717 -0.2661  1.6269 -1.6755 -0.0582 -0.1676
#> -0.0962 -0.5652 -0.5487  0.7782  0.1414 -0.8506  1.4646 -1.0331 -0.1788 -0.9083
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]