Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.4919 -0.5406 -1.3494 -1.3156  0.0344 -0.5670  0.0177 -0.1237  0.0032 -1.8809
#> -0.2353 -0.3880 -0.2357  1.7354 -2.3507 -0.7566 -0.4833  2.2890  0.6210 -0.5778
#> -0.3223 -0.5344 -1.2164  1.8265  0.6941 -0.4614  0.1875  0.2649 -0.4508  0.9058
#> -0.1063  0.2343 -0.3557  0.9641 -1.0679 -0.1649  0.5245  0.1400  0.5365 -0.1441
#>  0.1754  0.6344  1.1423 -0.4471 -1.0044  1.1475  0.4010  0.8862 -0.2278 -0.0741
#>  1.4150  1.3410 -0.7144 -0.1604  1.1226 -0.4764 -1.4031  1.2499 -2.3085 -2.0772
#>  0.5561  1.2233  0.9713 -0.3570  0.0178 -0.5985 -1.5756 -0.1410 -1.9229  0.3693
#> -0.3297 -0.1296 -2.1103  1.3779 -0.0620  0.1845 -1.9093 -0.3455  0.3412  0.5464
#> -0.7040  2.0245  1.5753  0.8795 -1.3282 -0.5002 -0.0812 -0.1888  0.3169  0.1737
#>  1.1656  0.4020 -1.3784  0.0510  0.2480  0.9042 -1.1907 -0.4088 -0.5119  0.5812
#>  0.2933  1.0301  0.2451 -0.2306 -0.0783  1.2610  1.3129 -0.6642  0.3777  0.6568
#>  0.1325 -0.3751  0.6844 -1.6802 -1.0547  0.6269 -1.2654  0.4893 -0.8965  1.1478
#> -0.0792 -0.0891  1.1853 -1.0799 -0.3445 -1.6559  0.0199  0.0399  0.5954 -0.0205
#>  0.3136 -0.3808  0.3305 -0.7252  1.4716  0.3789  0.7830  0.2508  0.2050  0.1597
#> -1.0706 -0.1220  0.8763 -0.5996 -0.1370 -1.4646 -1.4298  0.5136  0.1421 -1.1706
#> -0.1181  0.9562  0.6631  1.1482  0.1565 -0.4732  1.1789  1.4574  0.5758 -1.7255
#>  2.4703 -3.2096  1.6280  0.5984  0.0735  2.2546 -0.2638 -0.3146  0.4000  0.1614
#> -1.2404  0.1154 -0.0664  1.8806  1.6519  0.8452 -0.1900 -1.6052 -0.2373 -0.3576
#>  0.4821  0.6229  2.0793  0.4081  0.7247  0.0383 -1.5601 -1.4426 -0.2342 -0.1697
#>  1.2940 -0.6472  0.6638  0.3531 -0.0747  0.4157  0.2164  0.7791  0.3638  1.4015
#>  0.4052  0.0540  1.5016  1.7158  0.3658 -0.8953  0.8882  1.0805 -1.1932  3.0703
#>  1.1179 -0.2075 -0.6990  0.1607 -1.0441 -0.8542 -0.0849 -0.9984 -0.4356 -0.2284
#>  1.1134  0.6423  0.2389  0.2915 -0.4366  0.3149 -0.4916 -0.6608  0.9849 -0.3011
#> -1.3742 -1.0604  0.8233  0.3211  1.0405  0.2510  0.8128 -0.2359  0.6302 -0.3084
#> -0.2039  1.6210 -1.9645 -0.8072 -1.0294 -0.1447  1.3174 -0.1797 -1.4713  0.7767
#>  1.1360 -1.3724  1.6159  0.6726 -1.1375  0.0290 -0.0962 -0.3261  0.5070  0.1979
#> -0.5709 -1.5296  1.0830  0.2046 -0.5166 -0.0680  0.6239  0.0207  1.2004 -0.1524
#> -0.3085 -1.3510  0.0090 -1.6970 -1.6272  1.3688  1.6662  2.0683  0.0749 -0.8719
#>  0.2016 -0.3866  0.6392 -0.5225 -0.4884  0.7516 -1.1859 -1.0208  0.0031 -1.5555
#> -0.1142 -1.4745  0.7001  0.6450 -2.0296 -1.0703 -0.8443  0.5180  0.4139 -0.4592
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]