Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.4930 -0.3233  0.9027 -2.0744  0.1199  0.8967  1.3144  0.1240  1.5246 -0.3773
#> -1.0970  0.6310  1.7927 -0.5700  0.2936 -0.0518 -0.2851 -0.7750 -0.1427  0.0252
#>  2.4962  0.6230  0.5370  0.1725 -1.2882  0.9079 -1.1074 -0.0469 -0.5447  0.8910
#> -0.4907  0.4489  2.0610  0.7219 -0.6719 -0.1619  0.6287 -0.7037  0.9171 -2.0223
#>  0.3778 -0.4806  0.9575 -0.3137 -1.2677 -1.0720  1.0564 -0.7140 -0.9454  0.5860
#>  0.9446 -0.1961 -0.1397  1.1664 -0.3320 -1.0176 -1.0162 -0.1752  1.6442  0.2385
#> -0.5246 -1.7934 -0.9993  1.2912  0.5329  0.4195 -0.1021  0.4752 -1.3657  1.0437
#>  0.2159 -2.2961  0.4325  0.2422  0.4457  0.2057  0.6292  1.2386 -1.7295 -1.0723
#> -1.4281  0.6758 -0.8692 -1.5592 -0.6457  0.6472  0.6343  0.7767  0.5110 -1.5949
#> -0.3040 -1.1855 -1.6386 -0.5144  0.8429 -0.8824 -0.8067  0.4121  0.5115 -0.4074
#> -1.8287 -0.8693  0.2641 -1.3476  2.2211  1.9887 -0.2323 -0.4906 -1.3330 -0.9071
#>  0.2646  0.0563 -1.5005  0.0674 -0.8630 -0.5371  0.7442  1.4445 -1.0376 -0.8836
#>  0.0449  1.3548  1.2324 -0.5848 -0.7748 -0.0157 -0.8420  0.4436  0.6733  2.2461
#>  0.4477 -0.0851  0.3961  1.1305 -0.0113 -0.9732  0.2026 -0.6878  1.2368 -0.2926
#> -0.4702  1.1559 -1.1490 -0.3406  0.6689  1.9980 -0.1502  1.2575 -1.0630  1.3385
#> -0.3864 -1.3644  0.2797  0.4194  0.2121  0.8384 -0.3968 -0.6575  0.7516 -0.8950
#> -1.7290  0.2164  1.2101  1.5241 -0.2144  0.6128  0.1618  1.5275  1.0057  0.2857
#>  1.9954 -0.2113  0.8268 -0.7910  0.4053  0.4341  0.0679 -0.3361  0.1947 -1.7758
#> -0.8072  0.1083 -0.7156 -0.3553 -0.8746 -0.0711  1.0280  0.2482  1.3442 -1.1905
#>  1.0287 -0.2943  0.8204  0.3285  1.0812  0.0814  2.5043  0.8344  0.3093  0.8011
#>  1.4031  0.6488  0.1134 -0.7341  0.0155 -0.5816  1.0716  0.6146 -0.7224 -0.7224
#> -0.8945  0.5874 -0.0651  0.5785  1.2189 -1.3600 -2.6084  0.0170  0.9626  0.3881
#>  0.8202 -1.1941 -0.2977  0.4277 -0.7075 -1.0913 -0.6556 -0.0167 -1.5339 -0.2574
#> -0.2835 -1.0749 -0.1498 -1.0440 -0.3057  0.0412 -0.8785  0.6615  0.4340  0.1741
#> -0.3269 -0.4413  0.9617 -0.4364  0.1398 -0.5067  0.1716  1.2817  0.4443  1.5832
#> -0.8631  0.8086 -0.2574 -1.9954 -0.5166  1.1566 -0.4974 -0.2309  0.9245  1.6178
#>  0.5003 -0.7977 -0.1427  0.8978  0.0168  1.1793 -0.3319  1.4075 -1.9922 -0.7810
#> -0.2608  1.4676 -0.9970  0.3765 -0.8226 -1.0166  0.7136  0.3494  1.5008 -0.9507
#> -0.3230  0.1519 -1.6238 -0.3780  0.5759 -1.2360  2.3467  0.2532 -0.1293  0.8703
#> -0.0708  0.3788 -0.1492  0.5478  1.1496  0.7681  1.0292  0.6178 -0.9891  1.1157
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]