Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.8706 -0.7232 -0.2406 -0.6432  0.3620  0.0636 -1.6602 -0.1644  0.4481 -0.6861
#>  0.1474  1.4096  0.2876  0.8299 -0.9301  0.2693 -0.9142  1.5640 -0.3027 -0.3373
#>  0.0773  0.8570 -0.0359 -0.1220 -0.6155  0.4937  0.6775 -0.0111 -0.5300 -1.0904
#>  0.1905 -0.4250  0.6553 -2.1288 -0.9020  0.2002 -1.3535 -0.4450  0.2369 -1.5981
#> -0.5202  1.4336  2.5957 -0.8016 -0.0914  0.4388 -0.1089  1.0129 -0.3312  0.3934
#> -2.0638 -0.8441 -0.4584 -0.3326 -1.5577 -0.4705 -1.5961 -1.4546  0.3058  0.0897
#>  0.0844  0.9462 -0.5385 -0.7186 -1.1527  0.8062  0.1863  0.4399 -1.1034  0.8634
#>  0.4137 -0.3825  0.3010  0.7201  0.5096  0.3758 -0.4564  0.1791 -1.4084 -0.4600
#> -0.8951  1.3456 -0.4781 -0.4823  0.5523 -0.0630  0.0046  1.4045  2.2258 -1.8464
#> -0.0648  0.0980 -0.6859 -0.1529 -1.0860  0.3606 -0.8251  2.5755  0.8875 -0.0926
#>  1.4464 -0.0289  0.7880  0.6743 -0.0301 -0.8394 -0.0416 -1.5066  0.0850  0.5166
#> -0.9017 -0.1972  0.2535 -1.9976  1.5706  0.4100  0.3833 -0.2516  0.8718 -0.5629
#> -1.0487 -1.3302 -0.7393  0.4533  0.6782 -0.4516 -0.5600 -0.4475 -0.0421 -0.5592
#> -1.2212  0.6418  0.4276 -1.4193  0.1958  0.6362 -0.7087  2.7767 -1.1232 -0.2537
#> -1.0737 -0.3591  1.0507  1.5815  0.5985 -0.4512  0.7023 -1.6837  0.2106 -1.2200
#>  0.4234  1.5277 -0.3556 -0.6420  1.5556 -0.8215 -0.6400 -1.1232 -0.3843 -1.9862
#>  0.3853 -0.6189 -2.3332 -0.9623  0.8237 -0.1787  0.0780 -1.5793 -1.1824 -1.4320
#> -0.7602  0.1143  1.2725 -0.4375 -1.0207 -0.3133 -1.0960  1.0402 -1.2986  0.0459
#> -0.0321 -1.3808 -0.2345  1.4736  0.1957 -1.5171 -1.6100 -0.7911 -0.8255 -2.2509
#> -0.3460 -0.6107  0.2682 -0.0097 -0.3674 -2.1286  0.1026  1.6034  0.2280  1.5510
#>  0.8448  1.2353 -1.0151  2.1697  0.8019 -0.5840 -1.1306  0.1652  0.9584 -1.3299
#> -0.8529  1.2427 -0.3453 -0.0392 -1.5418  0.3002  0.1037 -0.7558  0.3781 -2.0184
#>  0.6147 -1.2080  0.6719  1.1396 -1.3231 -1.4335  2.9495 -0.0033  0.4285 -1.0186
#>  0.3497 -0.0343  0.0009 -0.2215 -1.6877  0.4791 -0.3381  0.7765  0.8553  0.0466
#>  0.2755  1.1399 -0.1303  1.3827 -0.0215 -2.0307  1.9421  0.0538  0.1012  0.3394
#>  0.0842  1.1345 -1.1492  0.3109 -0.3641  0.0394  1.1141  0.3182 -0.1395  0.0247
#> -0.9555 -0.3782  1.1395  0.7688  0.2950 -0.5431 -0.1250 -0.6980 -0.6695 -0.4980
#>  1.6940 -1.7978 -1.2469  1.4094 -0.4115 -1.2259 -0.4634  0.4700 -0.1051  0.1430
#> -2.1105 -0.7985  2.0543  1.1711  0.3894  0.0711  0.5293 -1.1842 -0.7297  0.9881
#> -0.7716  1.4548 -0.2171 -1.8418  0.6938  0.0541 -0.2849  0.8851 -0.7891 -1.4491
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]