Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.4705  0.2679  1.1328 -0.0978  0.3944 -1.0551 -0.7384  0.3530  0.2631  1.7950
#>  0.6877  2.0041 -1.6448  0.0380  1.2166  0.6304 -1.0763  0.9727  1.6971  0.6752
#> -1.0095 -0.4193  0.5037 -0.6566  0.6479 -0.5954  1.4334 -0.1087  0.2690 -0.4740
#>  1.1599  0.4402 -0.0199 -0.1429  0.6680  0.0181  1.4787 -0.0493  0.5409 -0.1888
#> -0.1874 -1.1435 -0.7365  0.7150  1.8496 -0.4704 -0.9842  0.3693  3.0963 -1.2800
#>  0.1488  0.9985  0.8469  0.0916 -1.0647 -0.2413 -0.9341 -0.4590  1.3408 -0.5234
#>  0.7550  1.5183 -0.4252 -1.5865  0.2246  1.3236  0.7886 -0.7493  0.0147  0.2340
#> -0.5279 -0.4579 -0.7218 -1.5686 -0.4330 -0.9814 -1.2333 -0.8225 -0.6683 -1.8454
#> -0.2662 -0.1670 -1.3876 -0.8720 -1.0105  0.1134  1.8705 -0.2700  0.9410 -0.6215
#>  2.3815 -0.1255  1.8345  1.3715  0.8324  0.0363  0.3920  0.4421  0.2133 -1.2351
#>  0.0829  0.3445  0.4191  1.0007  0.3772  0.9568  2.9860  0.2603 -0.9059  0.5708
#>  0.0321 -0.3067  0.8746  0.7198  0.7673 -1.4503  0.3663 -0.8408  0.4574 -0.1107
#> -1.1034  0.9801 -0.3932 -0.6743  0.2800 -0.0937 -0.7198 -0.7871 -0.7518 -1.0757
#>  3.1912  0.9836 -0.1201 -0.6997 -0.3732  0.8380  0.6531 -0.0118  0.3452 -0.4691
#>  1.0708 -0.8523  0.3654 -0.3510  1.8553  0.2474  1.0679  1.1412  0.3184  0.1563
#> -1.1815 -1.5485  0.2663 -0.3600  0.6363  0.1976 -0.5562 -0.8857  0.0157 -0.4466
#>  1.2452  0.4908  0.6836  1.7445  1.6333 -0.7969  1.0687 -0.5726  0.0485 -0.3978
#> -0.1459 -0.2935 -2.0460  0.2101  0.1138 -0.8258  0.2048  0.3262 -0.4472 -0.2437
#>  0.3720 -1.1158  1.0147 -1.4159 -0.1784  0.3512  0.3490 -1.1632  0.6055 -1.4777
#> -0.1135  0.1599  0.8635 -0.2293 -1.2565  0.5063  0.1240  0.6096 -0.0345 -0.4309
#> -0.3151  0.4923  1.1546 -0.3823  1.8475  0.4394 -2.1542  0.2684 -1.0709  0.4704
#>  1.4885  0.4509 -0.5420  2.0800 -1.3324  2.1283  0.6018 -1.3198 -1.8291 -1.4303
#> -0.5740 -0.7734  0.8168  1.0643  0.1760  0.7350 -0.7904 -1.2603 -0.4609  0.0763
#>  0.3399 -0.1658  0.6553 -0.7597 -0.2738 -0.8733  0.1550  0.5510  0.7544  1.3851
#>  1.7166 -1.4079  1.8896 -0.6904  0.9563 -0.6561  1.8740 -0.4865 -0.7584 -1.2846
#> -0.6491 -1.2459 -0.0122 -0.4750 -0.1256  0.8167  0.4127 -0.7807 -0.8320  0.8430
#>  2.0730 -1.0226 -0.3881 -0.2657  0.5991 -0.0465 -2.1225  1.3668  0.5399 -0.2199
#>  0.9357 -0.6902  1.1333  0.1227 -0.1665  0.2355 -0.4181  0.2707  0.3707 -0.3210
#>  0.7729  0.6190 -0.9693 -0.5309  0.6060 -0.5543 -0.5764 -0.9403 -0.0095  0.2895
#>  0.8740 -0.6915  0.0626 -2.5067  1.4603 -0.0084  1.9230  0.4045  2.2594  1.6072
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]