Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4964  0.0508 -1.3241 -0.5292  1.7141  0.6946 -1.5147  0.1290  1.4782 -0.0902
#>  1.6485 -1.5488  0.6674  1.6190 -0.4910 -0.8992  0.7487 -0.1699 -1.1910  0.8511
#>  1.3581  1.6708 -0.8260 -1.1250  1.0004  1.3074 -0.2946 -0.9906 -2.2215  0.8309
#> -0.0282  1.0685 -0.6579 -0.6325 -0.7902  1.6047  1.5362 -0.1546  0.0599  1.2308
#>  0.6731  0.5097 -1.8496 -0.9481  0.7004  1.1082  0.9037 -0.1458 -0.6227 -0.6227
#>  2.0177  0.4464 -1.0539 -0.3103  0.6479 -1.6629  0.3623 -1.4428 -0.0310  0.7463
#>  1.0952 -1.3624  0.3591  0.0686  0.7832 -1.2527  0.7155 -1.5103 -1.0258  1.4603
#> -0.0739  0.4309  0.1630  0.4375 -0.5375  0.7864 -1.3152  2.3967 -1.7234  1.3048
#>  0.1448 -0.6253  0.5395  0.1847 -0.4701  0.0180  0.5450 -1.7616 -2.1697  0.6698
#> -1.3677  1.8054  1.2293  0.9715  1.6067 -1.0657  0.9921 -0.3216  0.9031  0.8478
#>  1.0060  1.2171  0.0745  1.3360 -1.5122  1.1863 -0.7881 -0.1874 -0.7650 -0.4227
#> -0.5914  0.0971  0.1631 -1.1454 -0.5738 -0.9072 -0.1376 -0.0789 -0.8161  0.1804
#>  0.3911 -0.3395 -0.3176  0.7286  1.4966 -0.5957 -0.9611 -0.3235  1.1632 -1.4606
#>  1.3114  1.3247  0.1486  0.3415 -0.7845 -0.7814 -1.2543 -1.0442  0.4412  0.9376
#> -1.9850 -0.8512  0.6445 -0.5452  0.0335 -0.1675  0.6376 -1.5257 -0.0343 -0.5204
#>  1.0527 -1.1330 -0.5291 -1.0964  1.9608  0.5438 -0.5056  1.4202 -0.5640 -0.0820
#> -1.8755 -1.0444  1.1765  1.0947 -1.6870 -1.3744 -0.6812 -0.5650  0.1191 -1.8738
#>  0.4320  0.1831 -0.4502 -0.1112  0.0518 -1.1611 -1.2372  0.1830 -1.0184 -1.8905
#>  0.2018 -1.7422  0.7160 -0.1430 -0.9035 -0.3442 -0.1600 -1.9875  0.5114  0.1677
#>  1.5304 -0.3293 -2.3236  0.9875 -1.8303  0.6626  0.0075 -1.0722 -1.6837 -0.4305
#>  0.2200  0.3594 -0.3123 -0.8413  0.6005  0.1014 -0.5305  1.7214  0.3998 -0.6044
#> -0.3015 -1.0980 -1.8397  1.0186  0.4137 -1.0410 -1.4480 -0.3082 -0.8733 -1.7215
#> -0.4963 -1.3683  1.4041 -0.2839 -0.5615  0.3387 -1.5183  1.0840  0.2527  1.3465
#>  0.2281 -0.1514 -0.7405  0.4414 -0.1277  0.7463 -0.2126  0.1733 -2.0283  0.0681
#>  1.7761  1.7527 -0.0339 -0.3729  1.6401 -0.6466 -0.6586 -0.9891  0.2026  1.0401
#>  0.8491  0.8421 -0.3030 -0.6850  0.1906 -0.8498 -0.8454  0.2160 -0.4666 -1.4389
#> -0.0949  1.1364 -0.2604 -0.5748  0.1358  1.4755  1.5387 -0.1201  0.0368 -1.2270
#>  0.0991  0.1394  0.9532  0.0033  0.7899 -0.9642 -0.0410 -0.1605 -1.5582  0.8223
#> -0.2189  1.3402  0.6672  0.0384  1.3272 -1.4148  1.6613  1.5006 -0.4773  0.0987
#> -0.0950  1.4934  1.6179 -1.7178  0.3146  1.3577 -1.5812  0.4102  0.5701 -0.4794
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]