Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6-6.0470e-01 -4.1252e-01 -4.1031e-01 -9.5807e-01 -1.8933e+00 -2.7621e-02
#>  2.4678e-01  1.8568e-01 -1.7567e-01 -9.2057e-03 -5.0745e-01  1.6015e-01
#> -1.2049e+00 -3.8003e-01  4.6937e-01 -1.1061e-01 -7.0123e-01  9.1062e-01
#>  9.7521e-02 -4.0143e-01  1.0663e+00 -1.2812e+00  6.0722e-01 -1.1712e+00
#>  8.1571e-01 -1.2126e+00 -6.3665e-01  1.9309e-01  7.7182e-02 -7.1277e-01
#>  5.6840e-02 -2.1745e+00 -6.7031e-01 -1.2551e+00  1.6126e+00 -1.4307e+00
#> -1.8062e-01  1.3956e+00  1.7856e+00 -2.6394e-02 -1.8512e+00  9.3009e-01
#>  4.4991e-01 -1.0617e+00  2.9092e-01 -1.1815e+00 -9.4932e-01 -2.7066e-02
#>  7.3481e-01 -9.0469e-02  1.3666e+00  1.0148e+00  8.0054e-01 -2.4617e-01
#>  3.2626e-01  4.7564e-01 -2.0796e+00 -1.1899e+00  5.2495e-01  8.6280e-02
#>  2.2339e-01  4.0378e-01 -3.2131e-02  1.0890e+00 -1.6571e+00 -2.5595e-01
#> -1.3128e+00 -9.9444e-01  1.0966e+00  2.6563e-01  8.5195e-01  2.3407e-03
#> -5.0304e-01  2.3283e-01 -1.1472e+00  3.3134e-01  1.1543e+00  6.5518e-01
#>  1.5138e+00 -3.8393e-01 -3.1672e-01  3.6012e-01 -6.2197e-02 -4.5292e-01
#>  5.5462e-01 -1.2161e+00 -2.3467e-01  4.3001e-01  5.5199e-01  1.6800e+00
#> -1.8033e+00 -6.8954e-02 -5.8185e-01 -6.1165e-03  2.7863e-01 -2.1702e+00
#> -1.8617e+00 -1.5020e+00 -3.2154e-01 -1.4516e-01 -2.1828e+00  3.6044e-01
#> -2.3915e+00  5.3954e-01 -1.3905e+00 -2.7132e-01  3.7099e-01 -6.2359e-01
#> -2.4754e+00  1.3970e+00  1.0298e-01 -7.9525e-01 -8.9438e-01 -6.9475e-01
#> -2.4448e-01 -6.4609e-01  5.1086e-01  1.8050e-01  1.8031e+00  6.6797e-01
#> -5.3010e-01  1.0478e+00 -4.9742e-01 -5.9601e-01 -8.4845e-02 -2.0980e+00
#> -8.5113e-02 -4.0927e-01 -1.1916e+00  2.1017e-01  2.7525e-01 -7.6688e-02
#> -2.5464e-01 -2.7341e-01  1.6157e-01 -1.0206e+00 -2.1472e-01 -3.4481e+00
#> -6.4776e-01  4.9612e-01 -8.7165e-01  1.5330e-01  6.4300e-02 -4.8696e-01
#> -1.8900e+00  2.5400e-01  7.1435e-02  5.7727e-01 -8.1876e-01 -2.5556e+00
#>  2.0059e-01 -5.2990e-01 -1.8805e+00 -5.4676e-01 -6.2168e-01  7.2945e-01
#> -4.0451e-03  6.0386e-01  9.1457e-01  9.4974e-03  2.5476e+00 -8.7169e-01
#> -4.0926e-02  4.4290e-02 -8.5027e-02  9.4859e-01 -7.8432e-02 -1.7257e-01
#> -4.3752e-01  7.4365e-02 -1.0708e+00 -2.0735e+00  3.7775e-01 -9.0141e-02
#> -1.0455e+00 -6.9458e-01 -6.9341e-01 -4.5200e-01  4.0331e-01 -1.1438e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]