Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3057 -0.8800  0.9225 -0.6197 -0.7011 -0.5076 -0.3170 -0.3995  1.6858 -1.4504
#>  1.1333  2.6822  0.7160 -0.0916  0.8166  0.0612 -2.3341  0.0359  1.2745  1.4313
#> -1.6699  1.1148  0.4092  1.4523 -0.5042  0.3281 -1.5178  1.2359  1.5709  1.8425
#>  0.9103 -0.0140 -0.3646  1.0817 -1.3036 -0.6134  1.2455  1.0358 -0.1553  1.2991
#>  0.3418  0.2211  0.0548  0.8012  0.9807  0.2966 -0.2213 -2.3634 -1.7435 -0.8528
#> -1.0735  0.6857  1.0503  1.1922  0.1033 -0.7213  0.9187 -1.1139  2.2919  0.3797
#> -1.3143 -0.5802 -0.8932 -0.3244 -0.1703  0.4651 -0.1685 -0.5665  1.6617 -0.4395
#>  0.1303  0.1069 -2.4583  1.1308 -0.2916 -1.0204  0.8491  0.9819 -0.8731 -1.0196
#> -2.1454 -0.8926 -1.7719 -0.0691  0.3734 -0.4278  1.8667 -0.7242  0.2342  0.1192
#> -0.6834 -1.3626 -0.6368 -0.1153 -0.2682 -0.8211 -0.1302  0.8545  0.7348  0.1448
#>  0.9732  1.9758 -0.2607 -0.0874 -0.2374 -1.5317  0.6772 -0.4860  0.5611  0.8627
#> -0.5810 -1.3514  0.8504  1.4961 -0.7026 -0.4203 -1.4694 -0.6424 -0.6189 -0.6055
#>  1.7117  1.1058 -0.9319 -0.4516  1.1873 -0.2389 -0.0610  0.7840  0.1928 -1.1329
#>  0.3269 -0.3899  0.1299  1.7421 -0.9558 -0.1639 -0.2815  0.3840  0.2793 -0.1047
#> -0.8831  1.3620  0.3210  0.2407  1.3583 -1.2957 -0.1385  1.5291  0.1773  1.4786
#> -0.7638 -0.5333  2.2003  1.6291 -0.2993 -0.0163 -0.8682 -0.2634 -2.0048  1.8095
#> -0.0484 -0.4812  0.3213  0.0621  0.3771 -0.5428 -1.6523  0.3695  0.0323 -0.7710
#> -1.1837  0.5749  0.3357  0.6203  0.3761 -1.0349  0.6286 -1.5388 -0.6837 -0.4223
#>  0.2685  0.0819 -0.9008 -0.4965 -0.1157  0.3100 -0.4249  1.6997  0.7357  0.8096
#> -1.2576 -0.7308 -3.1579  0.2279  0.9453  0.1188  1.0739  0.2892  1.2141 -2.5109
#>  1.1844 -1.2852 -0.6360  0.7848  0.1600 -0.6612 -0.8592 -0.4914  0.2008 -0.0374
#>  1.5893 -1.4284 -0.7930 -0.0196 -0.0678 -0.9641  0.0351  0.7567  0.4748  0.6284
#> -1.5424 -0.6477 -0.9449 -0.1938 -0.4809  0.9612 -0.0131  0.5534 -0.5723  0.5828
#>  0.4685 -1.1081 -1.9000 -0.5699  0.6957  0.7213 -0.7955  0.3707  1.2182 -1.0798
#> -0.8188  0.7748 -1.0878 -0.2346 -0.4968  1.0795  0.5186  0.6121  1.0068  0.0770
#> -1.2322 -1.9815 -0.4202 -1.1507 -1.4981 -1.2997  0.3041 -0.2903  0.3034 -0.2666
#> -0.3870 -0.2177 -0.1471 -0.2311 -1.3259 -0.3816  0.4542 -0.9227  0.7409 -0.8995
#>  0.3762 -0.1839 -1.2339  2.0377  1.5563  0.3374 -1.0528  1.5828 -1.4628 -2.3705
#>  0.0272  0.3002  0.3393 -0.7844 -2.1947 -0.3903  0.8920  0.1794  0.3887  0.7194
#> -0.4290  0.2543 -0.7100 -0.6778 -0.0784 -0.2964 -0.2536  0.1450 -0.0698 -0.6894
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]