Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.4139  0.7894  1.6439 -0.4812  0.1098 -0.7416 -0.6167 -0.5487  0.5112  0.4507
#>  0.7475 -0.4086 -0.5169 -0.5088  0.7156 -0.9900 -0.2374 -0.5984  0.5059  0.8727
#> -0.6764 -0.3098  0.3438 -0.5757 -2.2436  1.8256  0.0024 -0.3988 -1.6835 -0.0779
#>  0.3633 -1.0374  0.1705 -0.7598  2.4009  0.6326 -0.5566 -1.0599  0.3923 -1.9224
#> -1.3460  0.1237  0.6116  0.2678 -0.9659 -0.4909 -0.0101 -1.0914 -0.1902 -0.2062
#>  1.0370 -0.4312  0.6502 -1.1587  1.3023  0.7312  2.6574  0.2522 -1.1546  1.8777
#>  0.6528 -0.9373  0.4063  0.7270 -0.6433 -0.6763  2.3044 -0.2467  0.4504  0.0447
#> -0.5698  2.3654 -0.1723 -1.0991  0.1152  0.0518 -0.4139  0.2356  0.8280  0.9868
#>  0.8111  0.1786 -1.8777 -0.1789  0.6688 -0.0501 -0.6652 -0.5111  1.3233 -0.5035
#>  0.3291  1.2957  0.2650 -0.9495 -1.4007 -0.7059 -0.8039 -0.2276  0.0306  1.4717
#>  1.0365 -0.5585 -0.4133  1.7699  0.8501 -0.0695  0.4481  1.7094  1.2293 -1.9316
#> -0.6135  1.0152  1.2810 -0.4369  1.7561  1.1521 -1.0627  1.0372 -1.0780 -0.2989
#> -1.5711 -0.0775 -1.3775 -0.0380  1.2307 -0.1629  0.8667  1.3274  0.3176  0.4925
#> -0.3330 -1.7792 -0.6114  1.5982  0.0642 -1.1742 -0.3452  1.4101 -1.6036 -1.1606
#> -3.2073 -1.6754  1.6087  0.0386 -1.2800  1.0994  0.4889  0.6308  0.0292 -0.4087
#> -0.3496 -1.1084 -0.2153  1.1310 -0.0003 -1.9379  0.8225 -2.7881 -0.0562  0.8672
#>  0.3250  0.3162  0.8673  1.1326 -0.4954  0.2925  0.6808 -0.4443 -1.6836 -0.2617
#>  0.2314 -1.3624 -0.6654  1.1399  0.2193 -1.7594  1.1109  2.4521  0.0137  0.9678
#> -0.6324  2.4537 -0.6748  0.3284 -1.2480  0.0258 -0.5894 -0.4494 -0.2192 -0.4947
#>  0.0411 -1.1373 -0.6190  1.4754  1.1023  0.1754 -0.7157  0.0891  0.6305 -0.6955
#>  0.9748 -1.4295  1.4353 -0.0782  0.6144 -0.0780 -0.1945  0.6087  0.6629 -0.4078
#> -1.2331  1.5005  0.6630 -0.1957 -0.4884  0.7512 -0.4814 -2.2352 -0.4902 -1.4397
#>  1.1998 -0.2043 -1.1880  0.0412 -0.1799  0.1346  0.6984  1.6452  1.7668  0.4535
#> -1.8784 -1.1419 -1.0849  1.7097 -1.5841  1.5666 -1.0563  0.3698  0.3315  0.1820
#>  0.7142  0.2317 -0.7461  0.4948  0.9001  0.9516  1.3789  1.3406  0.7717  0.0492
#> -0.7629 -1.1819  1.5303 -0.4941  0.4122 -0.3320 -0.7843  0.9948 -0.6004  1.1916
#> -0.2860  0.7976 -0.6210 -0.5661  0.0153  0.4551 -0.2646  1.1369 -1.3239 -0.3639
#>  1.5643  0.0614  0.8134 -0.6917  0.4914 -0.9927 -0.6182  0.6310 -1.5070 -0.6219
#> -0.0551 -0.9957 -0.1562  0.0478  0.8486  2.1182  2.1901 -0.5980 -1.8887 -1.2136
#> -0.0570  0.8195  0.8384 -0.4004 -0.8409  0.5027  0.1254  3.1604  0.1201  0.2779
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]