Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0465 -0.6457  0.8228 -0.5969  0.5174 -1.5096 -0.5270 -0.1182 -0.7168 -0.0494
#> -0.6539 -0.1557  0.5394 -0.4461 -1.2031  0.8914 -0.4058  0.8617  1.2313 -0.2976
#> -0.0699  0.0984  0.3873  0.2140  1.2179  0.9872  0.7059  0.5341 -0.5836 -0.0453
#> -3.1066  1.4056  0.8496 -1.1328  1.0465 -0.2827 -0.7590  0.1693 -0.7641 -0.8405
#> -2.6592 -1.2378  0.1876 -2.5852 -1.5422 -2.3570 -0.2591  0.4927  0.3548  1.1016
#>  1.2201  1.8677  0.4027 -0.0834 -0.2032  0.2147 -0.4949 -1.5331 -0.1993 -1.3302
#>  0.1238  0.8607  0.9396  0.2298  1.5133 -0.5634  2.1528  1.8540  0.5249 -1.4742
#>  0.6083 -0.2701 -0.5719  0.5146 -0.1153  1.0089  1.1063 -0.0276  1.8723 -0.3244
#>  0.6648  0.8858  1.2322 -0.8091  1.1304  0.2549 -0.0914 -0.6337 -1.2745 -0.4934
#>  0.0704 -1.2797 -0.2308  1.9922 -0.0789 -1.1706 -0.4547 -0.1392 -0.8427 -0.7872
#> -0.6979  0.5648 -1.0831  0.3200 -0.2302 -1.4409  0.6834 -2.1359  0.0635  0.9708
#> -0.8362 -1.8622  0.1492 -1.1027 -1.9176 -1.0635 -0.0318  0.8274  2.0420  0.5433
#> -1.0972  0.0261  0.8257 -0.3404 -0.4280 -0.5615  0.7305 -1.0731  0.3870  0.1636
#> -0.8426 -1.1986 -0.4613 -1.2582  0.5159 -0.4315  1.2717  1.1343  0.3611 -0.2600
#> -0.5110  0.2700 -0.3175 -1.4757 -1.0032  0.7343  0.7070  0.0586  0.2909  0.3539
#> -0.6121 -0.3079 -1.5992  0.7478  0.4137  0.4287 -1.1005 -1.4446  2.3029 -0.8707
#>  1.6563  0.4996  1.4451  1.4783  0.2323 -0.6433 -1.6307 -1.4578 -0.6522 -0.4721
#>  0.9998 -1.2880 -1.6715  0.4706 -0.1641  0.7515 -1.5146 -1.3437  0.4526 -0.1046
#>  0.3493  1.7792 -1.4478 -0.3326 -0.3049  1.8694  1.3069 -0.2373  2.1476 -0.6552
#> -0.1058  0.1119  0.4913  0.7353  0.7180  0.4319 -0.2393 -1.0131  0.7055  1.0792
#>  0.1606 -0.6656 -0.6502 -0.4101 -0.4399 -0.8159  0.1289 -0.0076  0.7093 -0.2148
#>  0.3421  0.1068  0.9726  1.2281 -0.9307  0.5785 -1.1014  1.8062  0.0039  0.7585
#> -1.0939 -1.2327 -1.1426  0.8853 -0.5753  1.7449 -0.6396  0.9469  0.8744  0.3548
#>  1.4164  1.1029 -1.2060 -1.5937  3.0876 -0.4468 -0.7217  1.4260  1.0948  1.0872
#> -0.8644 -0.5914 -0.4304 -1.0910  0.6911 -0.4309 -0.0376 -0.9726 -0.2884  0.0607
#> -0.3650  0.5453 -1.9105 -0.0091  0.1426  0.2151 -0.9840 -0.0544 -0.7963  1.0865
#> -0.1052 -0.9716 -1.6332  1.0644 -0.0754 -0.3843  0.9923 -0.9442  0.7714  1.1555
#>  0.6058 -1.1370 -0.7212 -0.7913  0.8572 -0.0339 -1.6263 -0.4618 -0.3966  0.3301
#>  0.0238 -0.1530  0.2190 -0.0131  1.3830  0.2645  0.2960 -1.1059 -0.9841 -0.6429
#>  0.8001 -0.3939  0.2424 -0.6046  0.7737 -0.3154 -0.5569 -0.3014  0.2968 -0.4713
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]