Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.2216 -1.0173 -0.0764 -0.2350 -1.9217  0.2746 -0.4885  0.0203 -0.0472 -1.4564
#>  1.0950  1.4944  0.3316  0.8373  1.4096  0.5286  1.0026 -0.4636 -0.0863 -0.6271
#>  0.6844  0.7347  0.0259 -2.9071 -1.4628  0.4867 -0.3887 -0.0312  0.9485 -0.5125
#>  0.3300  0.6403  0.0096  1.3864  0.0388  1.3447  1.2683  0.7262 -0.1932  2.0243
#>  0.3542  0.3241 -0.9187  0.7222 -1.7209  0.7460 -1.2555 -0.2333  1.5530 -0.6535
#>  0.1649 -0.3290  0.4064  0.7588  1.5105  0.6794  0.4745 -0.1727  1.3292  0.4091
#>  0.0504  1.4602 -0.1159 -0.0619 -1.9094 -0.5214  0.1625  0.0009 -0.0878  0.6175
#> -1.1037 -0.2119  0.4585 -0.9765 -1.5617 -0.0023  2.4902  0.3647  0.0815  0.2328
#> -0.0524 -0.9008 -1.3501 -0.1737 -0.3652  0.1730  0.0175  1.5803  0.1209 -0.5615
#>  0.9036 -0.1776 -1.4175  2.4191 -0.4502 -0.4245 -1.3338  0.1338 -1.3157  0.8370
#>  0.5184 -2.5591  0.2644 -0.1722 -0.8086 -0.6725 -0.5224  0.0108  0.0510 -0.0364
#> -0.7296  1.8346 -0.4295 -1.2537 -0.0495 -0.4311 -0.2022  0.4858 -1.2385  0.4426
#>  0.2876 -0.2874 -0.4098  0.2649 -0.7659 -0.7127 -0.4371 -1.2964  0.8682  1.8302
#>  0.5753 -2.1301  0.3901 -0.1430 -1.6707 -0.1899  0.9519 -0.0120 -1.8740 -0.7512
#>  0.4948  0.2577  0.3585  0.1585  2.1134 -0.1839  0.2365  0.6315 -0.7068  0.2929
#> -0.5917  2.3664 -0.8955 -3.1491 -0.4859 -1.2918 -0.8977 -0.3718  1.5527 -1.4609
#> -0.8542  1.6070  1.4402 -0.8703 -2.1383  0.0400 -0.5883  0.9617 -1.2811 -0.4916
#> -1.8845 -1.5787 -1.0694  1.7456 -0.1475 -0.6368 -1.9519  0.1361 -0.1988  0.8120
#> -0.5557  2.2575 -1.0858 -0.1631 -0.6413  2.0135  0.2119  0.4725 -1.8330  1.1276
#> -0.7864  0.3374  1.4867 -1.5838 -0.4317  0.7911 -1.3891  0.1107  0.0730 -0.5027
#>  0.3387 -1.4676  0.0002 -0.0260  0.2138  0.3348  0.7504  0.4215  0.3895  1.1705
#>  0.3794 -0.0082  0.2436  0.5953  2.6246 -0.9011 -0.3867  1.2334  0.0181  0.3534
#>  1.3819 -0.4995  0.8656  1.5840  0.5372  0.0247 -1.9571 -0.0412 -0.0524 -1.0627
#>  1.0895  0.3534  1.2270 -0.1487 -2.0640  0.0868 -0.0912  0.0736  1.0132  0.8696
#>  0.5385 -1.5120  0.4527 -0.1720  0.2609 -0.8788 -0.2486  0.7872 -1.9782  3.0195
#>  0.8033 -0.7350 -0.5300 -0.8177  0.6679  1.2552  0.3283  0.7141  0.1414 -0.2865
#> -0.7780  0.5551 -0.1710 -0.1793  0.8520  0.5529  1.1612 -0.5195 -2.0759  0.1461
#>  0.4063  1.3486  1.2169 -1.1040 -0.6827 -1.5075  0.2006 -0.7776  2.3024 -0.5206
#> -0.2149  0.5760 -0.3428 -0.8786  0.0914  0.5497  0.6302 -0.2663  1.2432 -1.2860
#> -1.8935 -1.5835 -0.3348  0.0953 -0.6331 -0.3628 -1.5677 -0.2168  0.0123 -1.7166
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]