Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6-1.8255e+00 -4.4634e-01  8.0532e-01 -5.7064e-01 -2.8991e-01  1.2585e+00
#> -6.9040e-01 -9.1996e-01 -5.3445e-01 -3.1270e-01 -7.9458e-01 -1.5706e+00
#> -1.0591e+00 -5.3751e-01  1.1680e+00  1.0306e+00  1.2699e+00 -1.5961e+00
#>  4.5650e-01  2.6850e-01  3.5387e-01  1.3113e-01 -4.2031e-01  2.4889e-01
#> -1.1271e-01 -1.8364e-01 -7.6105e-01  1.5299e+00  9.0703e-03 -9.4255e-01
#>  1.5019e+00 -1.8090e+00 -6.1227e-01 -8.0805e-01 -1.0280e+00 -4.9293e-01
#> -9.6597e-01 -4.9132e-01 -2.0609e+00  2.5890e+00  8.8270e-01 -6.6203e-01
#>  7.6950e-01 -2.1491e-01  7.4249e-01  1.2016e+00 -6.6785e-01 -1.2008e+00
#> -6.2179e-01  3.8731e-01 -1.0318e+00  3.4404e-01 -1.2668e+00 -2.0167e-01
#> -7.2657e-01 -1.4236e+00 -1.6175e+00 -1.1574e+00  5.2099e-01  1.0837e-01
#> -8.3840e-01  6.0789e-02  1.1170e+00 -9.6855e-01 -1.1535e+00  7.6570e-01
#>  3.4456e-01  1.2298e-01  8.0648e-01 -1.4550e+00  7.7238e-01  6.8317e-01
#> -3.5360e-01 -2.2086e+00 -1.2955e-01 -1.9389e+00  6.7824e-02 -1.6229e-01
#>  1.1091e+00 -8.0587e-02  9.2849e-02  1.9378e+00 -2.4915e-01 -7.9850e-01
#> -1.8487e+00  1.0165e-01 -5.6335e-01 -2.8140e-01  9.0376e-02 -5.9578e-01
#> -1.5345e+00 -1.1538e-01  8.8533e-01  1.4240e+00  3.2591e-01  1.2549e+00
#>  1.2335e+00  5.8575e-01 -1.6614e-01  4.4503e-01  5.6287e-01 -5.3605e-01
#>  1.1965e+00 -2.5745e+00 -2.7234e-01  1.3075e+00 -2.6265e-01  1.1818e+00
#>  8.1640e-01 -1.4185e+00 -1.1639e+00  1.1125e-01 -3.1259e-02  1.0367e+00
#> -8.7978e-01 -1.1062e+00 -1.3076e-01  3.8010e-01  8.9812e-02 -2.4856e-01
#> -9.8799e-01 -9.8680e-01 -4.6191e-01  3.1966e-01 -1.1753e+00  1.4899e-01
#> -9.7278e-01 -8.4412e-01  3.5439e-01 -8.4709e-01  1.8722e-01 -3.7577e-01
#> -4.5479e-01  1.1219e+00  6.0718e-01 -8.5149e-01  4.0040e-02 -6.9564e-01
#> -2.7752e-01 -1.7950e+00  5.0765e-02  2.5641e-01 -1.8838e+00  1.2386e+00
#>  2.3351e-02  1.3613e+00 -1.7737e+00  2.7030e-01  4.0578e-01  1.1005e+00
#> -1.1600e+00  3.5068e-01 -6.7117e-01  1.6939e+00  2.4120e+00  1.0748e+00
#>  9.6095e-01 -1.5708e-01  3.8096e-01 -9.8654e-01  8.1875e-01 -8.8394e-02
#> -6.5617e-01  2.5010e+00  6.6906e-01  7.5342e-01  8.0304e-01  2.0571e+00
#>  5.6545e-01 -2.3847e-01  1.9858e-01  6.9161e-01  1.4711e+00 -9.2644e-01
#> -1.9125e-01  1.5917e+00 -6.2658e-01  3.7176e-01 -8.4949e-01  7.2278e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]