Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.3954  2.3440 -0.9069  0.1831  1.7666 -0.3734  1.0155  0.8034  0.2179  1.6513
#>  0.4455 -0.8617 -0.2509  2.2568 -1.1322 -0.2356  0.8982  0.4967  0.0085 -0.5977
#>  0.0452  0.4817 -0.2691 -1.9456 -0.7338  0.2848  0.0038 -0.1282 -0.7337  1.4276
#>  0.9115 -1.1797 -0.0971 -0.3464  0.4613  0.3851  0.0277 -2.1199  0.8173  0.7418
#> -1.0591  1.5697  0.5600  0.0939  0.6203  0.5804  0.8835 -1.2842 -0.3438 -0.7777
#>  0.1081  0.9280 -0.9423 -0.9594 -0.5070  2.4794 -0.0510  1.4945  0.4792 -0.6710
#> -0.9298 -0.6401  1.2286  0.1033  1.0609  0.0417  0.3161  1.3328  0.6815  0.2685
#> -0.5030  1.4783 -0.2113  0.1948 -1.2481  0.2703  0.8232  0.3276  0.1204 -0.7304
#>  0.5312  1.7309  2.0251 -0.0266  0.4365 -0.2247  1.2332  0.0787  0.0492 -0.9333
#>  0.1155  1.1560  2.0037 -0.1597 -0.8715  0.9078  0.1080 -1.4604  0.0530 -1.7787
#> -0.9382 -0.8887  0.7082  0.8816 -1.0428  1.5895  0.3689  1.4872  0.4657 -0.0463
#>  2.5797 -1.5378  0.3940  0.3127  0.6015 -0.4203  0.1225  0.5265  0.2408  1.0124
#>  0.9048  0.9894  1.0070  0.2541  1.9289 -1.4927 -0.9824 -0.4670  1.1095 -0.3677
#> -0.2397 -0.2522 -0.1406 -0.7218 -0.5689 -0.1045 -0.4392 -0.3851 -0.4285  0.1112
#> -0.3628 -1.5420  0.4599 -0.1168 -0.7068  0.1007  0.1151  1.1276 -0.6737 -0.1330
#> -0.5086 -0.6774  0.5852 -0.3220 -0.8725  0.0392 -0.2049  0.5418 -0.2473  0.6261
#>  0.2263  0.1356  1.1970 -0.4182  0.2757 -1.3084 -1.3214 -0.2705  0.0959  1.6161
#>  0.8913  0.5335 -1.2299  0.8519  1.8504  0.0324 -0.3180  0.2642 -0.7078  0.7004
#>  0.8175  0.3570 -0.7658  0.7689  0.3021 -1.4526 -0.0889 -0.5938  0.6295  1.1255
#>  0.8069 -1.0392  1.4501  0.9135  1.8883  0.2817 -1.3978 -1.3515  1.2172  0.2574
#> -0.9201 -1.4053 -0.5038 -1.5803  0.2705  1.3837  0.7968  1.5763  0.5475 -0.0196
#> -3.2040  1.2110  2.1267 -0.9970  0.1736  1.3238 -0.1179  0.6064  0.9344 -0.3094
#>  1.4371 -0.2636 -0.0857 -1.1662 -2.0936  2.6733 -1.0754 -0.8346  0.0361  1.6741
#>  1.3125 -0.1601  1.5195 -0.0436  0.7967  0.2729 -1.1617  0.2417 -0.9298  2.0803
#>  1.0802 -0.0911 -1.2042 -0.4397 -0.2579  0.5529  0.5532  0.2417 -1.7443  1.2280
#>  0.2849  1.4634 -0.3918  0.9638 -0.9198 -0.4507  0.4704 -0.3404  0.0186 -0.1017
#>  0.9777 -0.6827  0.4220 -1.3107 -0.2063 -1.1061  1.6345  0.2321 -0.3073  2.4024
#>  0.5053  1.3125 -0.5667  0.2037 -1.0387 -0.3838  0.6493  0.9794  0.5580 -1.1507
#> -1.0554 -1.7667  0.7018  0.4525 -1.7522  0.1973 -0.4898 -0.4286 -1.4573 -0.2631
#>  1.3074 -0.4181  1.0175 -0.6085 -0.0776  1.5302 -1.2098  0.5403  0.5116 -1.0012
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]