Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7908 -0.1255  0.0271  1.6806 -1.1767 -2.4732 -0.5558 -0.6896 -0.5695 -0.1310
#> -1.2583 -1.3379 -0.4584  0.0484 -0.5049 -1.4832  0.2998 -0.0533  1.3858 -0.1767
#>  0.3731  0.3882  0.1688 -0.8348  0.0459  2.1045  0.9921 -0.8562 -0.7529 -0.9978
#>  0.8447  2.1119 -1.3971  1.1280  0.2851  0.0119  0.8937 -1.2615 -1.7887 -0.1593
#> -0.1543 -1.3809  0.2253 -0.6030  2.1223 -0.3821  0.9263 -0.4853  0.4169  1.1777
#>  0.4042 -0.1011  0.2667 -0.2842 -0.9187  1.6486  0.3013 -0.3833 -0.9033  1.0069
#>  1.0385  0.8633 -0.1987 -0.1495  0.0233 -0.4056  0.2606 -0.2284 -0.0292 -1.3037
#>  0.3081  2.3239 -0.4390 -1.0839 -1.4820 -0.6097 -0.0729  0.8994  0.2754  0.0526
#>  0.5243  0.6442  0.5362  0.2520  0.5142  0.3214  0.0583 -0.5561 -0.2604  3.3996
#> -0.5919 -1.4117  1.5831 -0.4288 -1.1209  1.7926  1.0841  0.5356 -0.3759  1.7843
#> -0.5957  0.4374  0.6989  0.9265  1.1627  0.6694  1.5434  2.6681 -0.2038  0.4783
#> -0.3597  0.2107  0.8661  1.3036  0.1344  0.0861  0.8645 -1.6559  0.2681  0.1747
#>  0.5256 -0.3002  0.6211  0.7142 -2.1047  0.8165  0.8401 -1.8489  0.8367  0.7573
#>  0.0640  0.4893 -0.5225 -0.6108  0.9514 -0.4194  0.6983  0.8164 -0.7808  0.1603
#>  0.0392 -0.6153  0.9188  0.2221  0.1551  0.2694 -0.5085 -0.5728 -0.6022 -1.0468
#> -0.0280 -0.3262 -0.7586 -0.0038 -0.1054  1.4316 -0.5087 -0.5622  0.0315  0.3918
#>  0.3113  2.4836  1.4658  1.4147 -0.0854  1.2682  0.4476 -0.1847  0.4251 -0.0902
#>  0.7008 -1.7569  0.8986 -0.8047 -0.2821 -0.5801 -0.1624  0.7727  0.0469 -0.3432
#> -0.6942  0.0469 -0.7340  0.0049  2.0873 -1.4945  0.1624 -0.1257  0.9121 -0.2985
#> -1.1137  1.2332 -0.7312  0.3904  0.6597 -0.6290 -0.8048 -2.2533  0.2213 -1.6475
#>  0.4571 -0.8643 -0.8132 -0.5688 -0.2963 -0.6543 -0.8371 -1.8668 -0.6697 -0.3404
#> -0.0587  0.7139 -0.3837  0.1636 -0.1715  0.6960  0.2382 -1.3353 -0.1584 -1.2934
#> -0.4001 -0.7671  1.0474 -0.7829  0.6474  1.0363 -0.2076 -2.1093  0.7521 -0.4653
#> -0.0675 -0.2573  0.0468 -0.1662 -0.9682  0.2028 -0.9869  1.6320  0.3658  0.0029
#> -0.1418 -0.5339  0.9734 -0.9966  1.4875 -0.3141  0.3313 -1.8099  0.3997  0.4276
#> -2.7894 -0.7823  0.7126  1.2138  0.6270 -0.6722 -0.2692 -0.3390 -0.2285  0.5363
#>  0.5941  0.6678 -0.7344 -1.5406  1.4951  0.8343  1.3809  1.0522  1.1105  0.3886
#> -1.9176  0.2447 -0.0728 -2.0020 -1.8498  0.5853 -0.2061  0.4902  1.0655 -0.0008
#> -0.3273  0.1137 -1.9118  0.4179  0.4997  0.3689  0.6331 -1.2703  1.4913  2.6303
#> -1.7092 -0.2303 -0.2838 -1.6104  0.6175 -1.0323  0.3624 -0.6589 -0.2366  0.4296
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]