Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.5314  1.2572 -0.1873 -0.4246  0.8389  0.6664  1.7306  1.2432  0.8374  1.2320
#>  0.3552  0.7925 -0.3207  0.5685  0.4647  0.6912  0.3295  0.8890 -1.2517 -1.0080
#> -0.8763 -0.3937  1.2656 -0.2420 -0.2186 -1.5454  0.5123  0.2788 -0.3477  0.4265
#>  1.2595 -1.0641  1.4761 -0.9112  0.5607  0.7651  2.6883 -0.4187  1.4589  3.0882
#>  0.5783  0.2371  0.5989 -0.3822  1.3167 -1.4193 -0.5848 -1.0961 -1.0633 -0.0072
#>  0.7676  0.1214 -0.8536 -2.1525 -1.1078 -1.5908 -0.2456  0.1052  1.0349  0.8484
#> -0.9521 -0.5660 -1.2011  0.9682 -0.9710 -0.5078 -0.6272  0.0462  1.1271  0.6105
#> -0.9085 -1.4676 -1.0236 -2.8360 -0.3161  0.5988  0.7125 -0.0234  0.9077 -1.5582
#>  0.0997  0.2840  0.9544 -0.2568 -0.4205 -1.2062  1.5924  1.5016 -0.8781  0.2237
#> -1.2425 -0.0796  0.5424 -1.7400 -0.5131 -1.8985  0.2775  0.2625  0.6986 -0.3246
#>  0.5363  1.0954  0.9117  1.3042 -1.5986 -0.7841  2.8346  0.6526 -2.0687 -0.4158
#> -0.9888  2.1958 -0.0355 -0.2388 -0.0944 -1.2036  1.3609  0.1860 -0.4055  0.0009
#>  0.3061 -0.1344  0.6857 -0.5682 -0.0566 -1.4997  1.4086 -1.8670  0.3156  1.8490
#>  0.5525  0.7375 -0.5576  0.4375  0.6438  0.0103 -0.0321  1.1217 -2.2904 -1.6364
#>  1.2255  0.7705  0.1560  1.1532 -0.1414  0.2076  0.3778 -1.2747 -0.0095  0.5543
#>  1.8297  0.4226 -0.0902 -1.6579  0.1234 -0.3294 -1.1731 -1.0095 -0.5163 -0.4532
#> -1.0537 -1.2082  1.3525  0.8201 -1.2015 -0.6291  0.9031 -0.2476 -1.4040  1.3681
#>  1.4066  0.3204 -0.1198 -1.7395  0.5724  0.8914  1.1068  1.1999 -1.7808 -0.4014
#> -0.1143  0.4065  0.1109 -0.5209  0.4003 -0.5999 -0.3097 -1.0145 -1.8954  0.8991
#>  0.1268  1.4260 -1.8313 -0.2919  0.5516  0.5205  1.7283 -1.1551  0.6368  0.7684
#>  1.8306 -1.2292 -0.7955 -0.2820 -0.6864 -1.1397  0.3681  0.6053 -0.5667  1.4713
#> -0.8054  0.0221 -0.7524 -0.9078 -0.3085 -0.0116  1.4673  0.2684  0.1835  1.7625
#>  0.0047  0.4009 -0.0047 -0.2966 -0.0397  0.1603 -2.6451 -0.7886 -0.1336  1.1016
#> -0.6879 -1.6978  1.5025 -2.0208  0.1203  0.7338 -0.7588  0.9576  0.3741  0.9296
#> -0.5593 -1.2138  2.4311 -1.9411 -0.6711 -0.6485 -1.5508  2.4816 -0.5064  0.6892
#>  1.0695  0.3467 -0.8516  0.8220 -0.8899 -0.3705 -0.5046  0.2381  0.1830  1.2750
#>  0.7997 -0.0707  0.3841  1.8292 -0.0466  0.3960 -0.2267 -1.0682  1.5855 -1.1219
#> -1.1209  1.3668 -0.5428 -1.2920  1.0513 -1.2835  0.0372 -0.2437 -1.8625 -1.0669
#> -2.7525  0.2382 -0.8978 -0.6697  0.7651 -1.0052 -0.5820  0.5517 -0.7926  0.8130
#>  0.0289 -0.4217 -0.1416  0.7698  0.2296 -0.9254 -0.0399  0.2550  0.4191  0.2391
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]