Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2488 -0.2940 -0.5467 -0.9106 -0.1565 -1.2814 -0.4355  0.0586 -0.2477 -0.9118
#>  1.1896  0.3418 -1.5058 -1.2840 -0.9474 -0.0425 -1.0322  1.1023  1.0381  0.3510
#> -1.0119  0.7214  0.0693 -0.4840  1.0973  0.6116 -1.2101 -0.9253 -0.8254  0.1683
#>  0.6320 -0.6155  0.5074 -1.8833  0.1013  0.2986 -0.4775  0.7643 -0.1331  1.2005
#> -0.9770  0.9617  0.2960 -0.3016 -0.0242 -0.7122 -0.4061  0.5555  0.1242  0.9887
#> -1.3891  0.5664 -0.1344  0.5677  0.0873  1.6861 -0.2251  0.8359  0.9901 -0.4796
#>  0.9241  2.0603 -0.9063 -0.2276  1.3867 -0.9341  0.5982 -0.1824 -0.6376 -1.1484
#>  0.4061 -0.6275  0.2707  0.4903 -0.1043 -1.9632 -0.0016 -0.7912 -1.1765  0.0613
#> -0.4257 -0.3930 -0.3285  1.3767  1.7935 -0.3119 -0.8934 -1.0842 -1.6394 -2.1890
#>  0.7970 -0.8066  1.4877  0.9026 -2.0513  0.1153 -0.1206  0.1854  0.0023  1.3901
#>  0.5397 -1.8536  0.2748  0.0114  0.6575  0.0231 -1.0483  1.1801  0.4484  0.1210
#>  0.0155 -0.1470  0.9055  0.4400  1.4924  0.8300 -0.6105 -0.0822  0.5883 -0.9488
#> -0.1815 -1.5037 -1.6271  0.9785  0.3817  0.5406 -0.2294  0.1728 -0.2700 -0.5407
#>  0.0841  0.2982 -0.4775  0.6508  1.1096  0.7870 -0.3857  1.2791 -2.8228  0.4046
#> -0.0869  1.3054 -0.2538 -0.1918  0.4946 -0.0159 -1.3590  0.8148  0.0953 -0.0854
#> -0.6926 -1.1296  0.4809  1.5120 -0.7866 -0.9310 -1.1357  0.0290  0.2656  1.2693
#>  0.6480 -0.6057 -0.2625  1.9982  0.0130 -0.7231 -0.3261  0.7078 -0.4379 -0.8238
#>  0.3505  0.2880 -0.1751  1.1721  0.9702 -0.3210 -1.0824  0.7465 -1.2222  1.8251
#>  1.8212  0.2372 -1.2343  0.2880  0.7393  0.8233  0.8004 -0.8841  0.7446 -1.2424
#>  0.5056 -0.6684  0.7946 -0.1071  1.2244 -0.1224  1.2629 -0.6903  0.9706 -0.8781
#> -0.9384 -0.2684 -0.1225 -1.1061  0.1277  1.1793 -1.4873 -1.8267 -1.2761  0.5361
#>  0.0299 -0.9193 -2.2762  0.5891  0.6558  2.0045 -0.0667 -0.5525 -0.7339 -0.0307
#> -2.3852 -0.0615 -1.4649  1.5499  0.1635  0.0045  0.6328 -1.0043 -1.0478  2.2857
#> -0.3420 -0.9398 -0.9311 -0.1491 -1.2075  0.2935  1.2260 -0.4335 -0.0549 -0.6587
#> -1.6248 -1.3434 -1.0498  0.6214 -1.6143  0.0368  0.4602  0.7831 -0.9098 -0.1874
#> -0.5805  0.2101 -1.5259 -0.1524  0.1000  0.0783  0.2097 -1.1822 -1.0175  0.7746
#> -1.0502  1.4101 -0.4055  0.6232  0.6228 -2.7328 -2.0389 -2.7017  1.6964  1.0835
#> -0.1121 -0.7196  0.5120  0.8707  0.8995 -0.6605  0.4494 -1.9254  1.6200 -2.0138
#> -1.6336 -0.2950  1.2028  0.8946 -0.0221 -0.0973  1.1457 -1.3285 -0.5265 -1.2223
#>  0.8526  0.6998 -0.8892  1.0884 -1.2780  0.1840  0.6539  0.6941 -0.6917 -1.3040
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]