Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.7366  0.4133 -0.1473 -1.1735  0.3395  0.5958 -0.2629 -0.0903 -0.0716 -0.7767
#>  0.8752  0.3963  0.3456  1.3134  0.3506  0.6830  0.2301  1.4630  0.9744  0.2942
#> -2.2753 -1.3081 -0.9647  0.0690  0.0427 -0.8771 -0.0089 -0.1946  0.0585 -0.3632
#> -1.1589  0.3112 -0.1348 -0.8210 -0.6773  2.6438 -0.3429 -0.8200 -0.1614  2.2701
#>  0.0356  0.7118  0.4878  1.6333 -1.0709 -1.0435  0.0454 -2.3155 -0.0399 -1.3227
#>  1.3408 -1.4536  1.8222 -0.1622 -1.0452  0.1701 -0.0515 -0.4697  0.4055 -0.2705
#>  0.2413  0.0577  0.1711  1.6457  1.9059 -1.2155 -1.5303  0.8112  1.2505  2.0421
#>  1.0701 -1.8206 -0.6330 -0.3567  2.6253  0.7269  1.5258  0.8329 -0.7729  2.0090
#> -0.3396  1.1973  0.3289  0.2058  0.1522  1.4237 -0.0865 -0.5986 -1.9463  0.8103
#>  0.7231 -0.3169 -0.2821  0.6758  1.1123 -0.9878  1.1915 -1.0264  0.6137  0.8296
#> -0.1208 -0.9103 -1.4980 -0.2402 -0.5232 -1.8305 -0.3468 -0.2952 -0.3540 -1.2111
#>  0.0223  1.0870  2.5036  0.7422  0.4812  1.1825  1.5183  0.0967  1.2121 -0.2620
#> -0.1853 -0.8530  1.0646  0.7902 -0.0502 -0.1617  0.3496  1.3770  0.2285  0.6012
#> -0.3186 -2.0000  0.5311  0.4601  0.7466  0.4836  0.6608  1.7294 -0.0473 -0.4518
#>  0.2531 -1.0987  0.1382 -0.9209 -0.0272 -0.0287  1.4842 -0.8005  0.0961  1.1477
#>  0.6607 -1.1038  1.4098  1.1243  0.7404 -1.8365  0.7864 -0.8400 -0.6970  0.9643
#>  0.1288  0.0959 -0.2289  1.5315  2.3234 -0.3444 -0.4005 -0.1069  0.2905  0.7343
#>  0.3515 -0.4947 -0.4069  1.3436  1.2474 -1.0420 -1.3099  1.1215  0.5877  0.2140
#>  1.2089 -0.4521  0.2125 -0.2855  0.5461 -0.1470 -1.4117  1.6280 -0.0893 -0.9536
#> -1.1686 -0.8077  0.3438  0.2991 -0.0246 -0.2495  0.5050 -0.8141 -0.6723  1.4398
#>  0.7258  0.0213 -0.9083  0.6185  1.1216  1.6186  1.6232  0.6398 -0.3960 -0.9356
#>  1.2263  0.7374 -0.5994  0.7500 -0.0644 -0.5100 -1.2788 -1.4780 -0.6941  0.7573
#>  0.9348 -0.8037 -1.7531  2.5875  0.3416  1.7174  0.7533 -0.1895  0.0281  1.9230
#>  1.3901 -0.6096 -0.6543 -0.7440 -1.4885 -0.3611 -0.9410  0.8287 -1.5255  1.5104
#> -0.3467 -1.7252 -1.9877 -0.6056 -0.3029 -1.6590 -0.2590 -1.0963 -0.1401 -0.1108
#>  0.7479 -0.0714  0.2240  0.4749 -1.7861 -1.1205  0.3939  0.6712 -0.4615  0.9124
#>  1.5201  0.0249 -1.5585 -1.4093  0.3011 -1.2725 -0.7776 -1.8006  0.6567  0.1255
#> -1.1761 -0.4981  0.7644 -3.1858  0.7507  0.2787 -0.9768 -1.7335 -0.0086 -0.0634
#>  1.3970  2.0750  2.1602  1.2978 -0.6330  0.7955  0.1989  0.0649  0.5070 -1.3999
#>  0.0264 -0.1664  0.9490 -0.8003 -1.4767  0.9042  0.7181 -1.3780 -0.3432  1.3388
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]