Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3721  0.9092 -0.8785  1.9468 -0.3256  3.6324 -2.2698 -0.5340  0.6440 -1.2292
#> -2.1218 -0.8265 -0.8676 -0.9825 -1.8462  0.7472  1.0661  1.0017  0.3423 -1.0399
#>  0.3185  1.0992  0.9777  0.3030 -1.2179  0.6039 -0.2080 -0.0350  0.0328  0.2922
#> -0.2515  1.2229 -0.7374  1.0704  0.0376  0.6830 -0.1156  0.1825 -0.0132 -2.0318
#> -0.4790 -0.4979 -0.6367  0.7717 -1.1110 -0.6971 -0.7174 -1.2584  0.1195 -0.5691
#> -0.3401 -0.2882  0.4530 -0.3667  1.2741 -0.3571 -0.8975 -0.2633  0.6724 -0.1908
#> -0.3935  1.6458  0.3519 -1.1618 -0.7240  0.9256 -0.0026 -0.1268 -1.8616  0.2950
#> -0.8119  1.7136 -0.9078  0.7501  0.0284 -0.1353 -0.0038 -1.2541  1.2324 -0.2824
#>  1.3031 -0.1935  0.0818  0.5056 -0.3432  0.1825  1.4444  0.6907 -0.0965 -1.8855
#>  1.4547 -1.1775 -0.5664 -1.4115 -0.3583  0.4310 -0.0469  0.9127  0.1009  0.2705
#> -0.3114  0.9463 -0.7418  0.0077 -0.3569  1.1442  0.9262 -0.9340  0.3432 -1.5502
#> -0.4975  0.1114 -1.2202  1.9255 -0.1047 -0.7216 -0.1468  0.7524  0.4075  0.3540
#> -1.0566  1.0500 -0.6338 -0.7007 -1.3424 -0.4931 -0.0570  0.1071 -0.0937  0.3427
#>  0.2064  0.8642  0.1282  1.0504 -1.0989  0.4665  0.1071 -1.2311  0.6726  0.5570
#> -0.6736  0.8500 -0.1266 -0.5957 -1.0345 -0.3181  1.5714 -0.3161  2.0944  0.4883
#>  0.5677  0.1064 -0.2422  0.7261 -0.7807 -1.1571 -1.5996 -0.2006 -1.0894  0.9928
#>  0.3503  0.1395  0.4212 -0.5934  1.1779  0.0413 -1.1193  0.2768  0.7870  2.0835
#>  0.0056  1.3875  1.4146  0.1654  0.7684 -1.7285 -0.7445 -0.7779  1.7151 -1.1599
#> -0.9795 -0.8390 -0.6107 -0.2301  0.6500  0.0178 -0.5232 -0.7402  1.1688 -0.3203
#>  0.4737  1.4313  1.5952  0.0698 -1.8069 -0.6003 -1.8412  0.9226 -0.9157 -0.0382
#> -0.4900 -1.0681  0.0099  2.3986  0.7091  0.0979  0.2984  0.0041 -0.6417 -1.7353
#> -0.0479 -1.4799  1.7156  0.1403  0.7425 -1.2987  0.5475  0.1620 -1.7447  0.9525
#> -1.0271  0.2022  2.8100 -1.1392 -2.0947 -0.3420  1.8146  0.5973 -0.0976 -0.0616
#> -0.0633  0.3751 -0.7388  2.1898 -0.4665 -0.6890 -0.5783 -0.0096 -2.5355  0.7142
#>  0.2488 -0.4053  0.3248 -0.1260  0.9934  0.7347 -0.9229 -0.3677 -0.9752 -1.2007
#>  0.7723  1.9969 -1.2736 -0.2481  1.1267  1.3981  0.4029  0.4168 -0.2823  0.5721
#>  0.0493  0.5381  1.2854 -0.5371 -0.3750 -0.4882  1.3138  0.2180  0.5198 -0.9049
#> -0.9405 -1.3058  0.1339 -0.1259 -0.6627  0.8461  2.4109 -0.3074  1.5352 -0.3390
#> -0.3415 -0.1254 -0.2867  0.1995  2.0854 -0.9578  1.1084 -0.0724 -0.6947 -0.3444
#> -0.6925  0.0201  1.4316  0.7606 -0.2947  0.1095 -1.0753  1.1575  1.9056 -0.0901
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]