Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-3.2001 -0.8096 -0.4241 -0.8921 -1.2799  1.2628  0.0614  0.3115 -0.0076 -1.1838
#>  1.6840 -0.2099  0.1236 -0.3638 -0.7913 -0.2537 -0.4428  0.0821 -0.6477 -0.3585
#>  0.2320 -1.4370  1.2340 -0.8679 -0.6273 -0.3855 -0.0800 -0.2398 -0.0473  1.5308
#> -0.8860 -0.4551  0.2983 -1.7410  1.3985 -1.3567  0.8211  0.5397 -0.3540  0.0272
#>  1.3797  0.1674 -1.0552 -0.7849 -1.2433 -1.2122 -1.3759 -0.1765 -0.0273 -1.5122
#>  0.1401 -2.1685  0.7897  1.0729  1.6522 -0.1352 -1.3181  0.0045  0.3579  0.8393
#> -1.5310 -2.5207 -0.8466  0.3493  0.5615 -0.2642 -0.1810 -1.0460 -0.2123 -0.2413
#> -1.0873  0.1273  2.1211  1.6426  1.4075  0.6777 -1.6513 -0.3243  0.3490  0.3645
#> -1.7088 -2.1270 -0.1810  1.4104  1.8593 -0.9127 -1.2981  0.4698 -0.1992 -0.1140
#>  1.9378  0.9732 -0.0236 -1.1482  1.4814  0.1334  1.1554  1.3682  1.6263 -0.7402
#>  0.5368 -0.0813  1.1991  0.1924 -0.6889  0.7345  1.2382 -1.7861  0.5173 -3.0856
#> -0.7968  0.1690 -0.7993  0.1750 -2.1596 -0.6042 -0.4548  1.0959  0.8734 -0.9513
#>  0.9880 -0.3118  2.1232  1.6591  1.2849  0.9249  0.6619  2.0033  0.1736  0.5763
#> -2.0815 -1.8440 -1.5291 -1.6997 -1.6071  0.6499 -1.1154 -0.9334 -0.6603 -1.0174
#>  1.0461  0.9842 -2.5651 -0.1731  1.1544  0.0017 -0.1712  0.2549  0.6454 -0.2433
#>  0.6854  0.4113 -0.7559  0.5953  0.8224  0.2375  0.1328 -0.5328 -1.1350 -0.7388
#>  0.9265 -2.1994  0.7919  0.4563 -0.5612 -0.4017  0.5140  0.4877  0.1908 -1.0547
#> -0.0882 -0.4833  1.2140 -0.2970 -1.6036 -1.2172  0.4129  0.1495  0.2382 -1.8553
#>  0.0580  0.4877 -0.0815  0.0335 -0.7675 -0.0857  1.0799 -0.7747  0.8888  0.4437
#>  0.7695  2.1487  0.7349  0.8489 -2.4781  1.1257 -1.2582 -0.8949  0.1064  0.7861
#> -0.1021 -0.5930  0.6860 -0.3893  0.5605 -1.2665 -0.3630  0.9917  1.1378 -0.2819
#>  1.6451 -0.9201  0.3021  0.3005  0.7206 -0.9413 -0.2403  0.0708  0.2361 -0.8043
#> -1.6057  0.6335  0.2466 -0.6638  1.0764 -0.3837  0.0804  0.6517 -0.0191 -0.0863
#> -2.6790  0.1633  0.2949  0.5391  0.5722 -0.0427 -0.0619  1.0400  0.0698  0.0995
#> -0.6940  0.1729 -0.8914  0.4354 -0.3233  2.7008  1.5827 -0.0310 -0.2023 -3.4834
#>  0.1199 -0.5034 -0.5505  0.6626 -0.1511 -0.0548  1.8774  0.0196  0.3651  3.0259
#> -1.6961 -1.5236  1.4559 -1.2928 -0.8739  0.1066 -1.1214  0.0610 -0.1703 -0.4621
#> -0.9059 -0.4414 -0.8729  0.1922 -0.6445  0.5570  0.3422  1.4164 -0.4022  1.2881
#>  1.4514 -2.3132 -0.1349 -2.0758 -0.4972 -1.0275  2.0469 -0.3473 -0.3170  0.1763
#> -0.5431 -0.6586  0.3006  1.5637 -0.7207 -0.1404  0.6172 -0.1942 -0.3028 -0.5314
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]