Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3169 -0.1862  0.6522 -1.0848 -0.0624 -1.2537  1.7205  0.4170  0.0288  0.1189
#>  1.2341  0.1650 -1.0467 -0.7195 -0.5055 -0.3974 -0.2246 -1.8533 -0.0990  0.3433
#> -0.4467  1.9918 -1.4496 -0.7155 -1.6798  0.7947 -1.1604 -0.1578 -0.4409  0.5786
#> -1.2795 -0.3431  0.0211  0.6676  0.3115 -1.5485  0.5792  0.7865 -0.4771 -1.3046
#>  0.4958  0.3624  1.2993 -0.6869  0.7682  0.1333 -1.2049  0.1000 -0.2068  0.0524
#>  0.5674 -1.8741  0.1652 -0.5555 -0.7343 -0.5317  0.7117  0.3399  0.3408  1.1173
#>  0.4456 -1.6200  0.8523  1.9573  0.5712 -1.1306 -0.6665  1.8706  2.4954  0.9862
#>  0.6608  1.5139 -0.2575 -0.8343  1.4522 -1.0437 -0.1102  0.0963 -0.0737  0.6301
#> -0.1299  2.1558 -1.2802 -0.0476  0.0584  0.4346  1.3229 -0.0084  0.4182  1.7118
#> -0.2670 -0.7795 -0.9894  1.7673  0.0651 -0.6101 -0.1591  1.1759  0.3979 -0.7903
#>  0.5993  0.1740  0.6875 -0.2064  0.2802  0.7443 -0.2857 -1.0052 -0.0314 -1.6907
#> -0.8277 -0.4096 -0.8941  1.1567 -0.6118 -0.3159  0.6613 -1.2076 -0.3889  0.5743
#> -0.9857  0.1477 -1.2625  2.1227  0.2904  2.6333  1.0004  0.0295 -1.4524 -0.3493
#>  0.9469  0.2207  0.4452  2.4175 -0.2516 -1.2511  0.5766 -0.3307 -1.6740  0.3054
#> -0.5583 -0.4995 -0.7163  0.5807 -0.4468 -0.1272  0.4831  0.6458 -0.5647 -0.7729
#> -1.7469  1.1134 -0.2439 -0.7310 -0.1957 -0.6561  1.0817 -1.0407 -1.6228  1.1563
#> -0.7806  0.0141  0.6557  0.3411 -0.4467  0.0236 -1.0491  0.0994  0.3537 -1.3623
#>  0.6561 -1.5371  0.7739  1.5565 -0.1508  0.2087 -0.4348 -0.2235 -2.1854 -1.5594
#>  0.7137  0.1547  0.6393  1.4791  0.1341 -0.0025 -0.0355  1.9509 -0.2472  0.7331
#>  0.6299  1.6360  1.3042 -0.1250 -0.3053  2.5411 -0.4233 -0.6612  0.1490 -1.2732
#> -1.7969  0.0327  0.2213 -0.4256  0.2667  2.8970  2.2442 -0.3468 -0.4740 -1.9675
#> -0.9390  0.5952 -0.4921 -0.7823  0.4025  0.7044 -0.0824  1.0612 -0.2431 -1.4253
#> -0.8389 -1.1099  1.9692  0.6792  0.6221  0.3749  0.1043 -1.6522 -0.6438  1.8644
#>  0.6379  2.2797  1.9191 -0.8191 -1.0407  1.4876  1.0285  0.3960 -0.5382  0.5758
#>  0.4569  0.9190 -0.8264 -0.9113  0.9284  1.3230  0.8700 -0.4956  0.7767 -0.8385
#> -1.4130 -2.1864  0.7006  0.9804  1.4527  0.0515  0.6414 -1.4515 -1.0128 -0.1474
#>  0.3322 -1.6724 -2.2710 -0.6021 -0.0135 -0.0140 -1.0582 -0.0202  0.0186 -1.5592
#> -0.1542 -0.2052 -1.0222 -0.8335  0.7901  0.4980  0.5313 -1.4186  0.0194 -0.3471
#>  0.3451  0.4855 -0.7451 -0.0294 -0.9146 -0.6036 -1.1116  0.7432  1.9569 -1.1001
#>  0.3710 -0.0337  1.0160  0.1134  1.2934 -1.0365 -0.4428  0.5203  0.2367 -1.6886
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]