Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2711  0.5554 -0.3673  0.1534 -0.2643 -0.7284  0.8238 -1.7668  1.2938  0.6379
#> -0.7590  0.5741  1.7266 -1.1280  1.1102  0.0724  0.1975 -0.6771 -0.5444  1.5028
#>  1.5433 -0.3824 -1.6942 -0.1134  0.8947 -1.8734  0.4682  1.5929 -1.1189  0.7359
#> -0.7395 -0.4168 -1.0306 -0.0368  1.3284 -0.2815 -0.1735  0.9471  1.4622 -0.0799
#> -0.8803 -1.3612 -0.4682 -0.9527  0.0502 -1.3725  1.0518  1.8199  1.1728 -0.0462
#>  1.1569 -0.7309 -0.3363 -0.3382  0.2431  0.7992 -0.3513  0.2943 -0.6125 -0.4140
#>  0.3853 -1.0053 -0.4475  1.3469  0.3332  0.2461  1.6665 -0.9964  0.8598  1.9046
#> -0.5348  0.5682  1.9431  1.0687 -1.0374  1.1230 -0.5968  0.1006 -1.2037 -0.7609
#> -1.1160 -1.8947  0.3439 -0.6852  0.0933 -2.3857 -0.1996  1.3371 -0.4824 -1.1262
#>  0.7358  0.0256 -0.1853 -0.5476  0.0630  0.4678  0.8114 -0.2432 -1.5922  0.3386
#> -0.5027  0.2825 -0.9874 -0.3359  0.1722 -0.6246  0.6443  0.5315 -0.2696 -0.6930
#>  0.3758  1.5848  0.3622 -1.8574  0.2526  1.0579  1.8431  0.3301 -0.3274 -0.1684
#> -0.5124  1.1245 -1.1127 -0.2441 -0.2843 -0.9682  0.2242 -0.1214  0.5585  0.0100
#> -1.3319 -0.2650 -0.3214 -0.1712 -0.0816  1.8588 -1.6097  1.7250  0.4777 -0.6810
#> -1.0109 -0.4163  0.9246 -0.6202  1.7525 -0.7211  1.2037  0.4040 -0.2845  1.0928
#>  0.8113  1.6518  0.5924 -1.0277  1.0220  0.5627 -0.8794 -2.1835  1.2247 -0.8522
#>  1.1900  0.2043  0.0458  1.5208  0.4893 -0.0431  1.5079  0.6600  0.8951  0.3483
#>  0.1625  0.8154  0.9012 -0.9912  1.0319 -1.1297 -0.9656  1.5332  0.2587 -0.7533
#> -0.3038 -0.6507 -0.6988  1.1514 -0.2516  0.3018  0.8942 -1.4886 -0.5909  0.6451
#>  1.2356 -1.4652 -0.1788  0.7292  0.2064 -0.8670 -0.3424  0.4071 -1.5958  0.2036
#> -0.2692  1.4955 -0.8464  0.9404 -1.6658  0.1316  0.9266 -1.3291  0.0319  1.8233
#> -0.4589 -0.5512  0.5762 -1.0948  0.8090 -0.2183  0.4030 -0.1519  1.1395  0.5486
#> -0.9855 -0.6130  0.4399 -0.7377  0.5834  0.0831 -1.3178 -1.3470 -1.5919  0.7334
#> -0.3433  1.0068  1.7588 -0.6315 -0.7374 -0.0099 -0.5333 -0.1852  1.9359 -1.3104
#> -1.1799 -0.6444  0.4676 -1.1942 -0.7341  0.1296 -0.9933  0.0909  0.4444  0.7227
#> -1.4075 -1.3522 -1.2286  1.6970  0.4513  0.5511 -1.8375  0.6755  0.3769 -0.3233
#>  1.0242 -0.3352 -0.7192 -1.1839  0.0062  0.5395 -0.8267 -0.2677 -1.1085  2.0452
#>  1.4725 -0.5405  2.2585  0.5371  0.1309 -1.5170  0.1724  0.7129  1.8068 -1.1821
#>  1.3480  0.4635 -0.9738 -0.5911  0.1537 -0.8244  0.7231 -1.1877 -1.3225 -1.2016
#>  0.8179  0.3926  0.6547  1.3459  0.2547 -1.1316  0.7631  1.1948  0.5519 -0.2914
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]