Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 2.2037  0.2691 -2.4223 -1.0899 -0.3955  0.0414 -0.8575 -0.6590  0.7855 -0.0919
#>  1.3091  0.5248  1.8762  1.6433  0.2890  0.8956 -2.5413 -1.6546 -0.6791 -0.9311
#>  0.6536 -0.4839  0.1773 -1.0948  2.4655  0.7559 -1.7078  0.3015 -0.3309 -0.5268
#>  1.7935 -0.4728 -0.2556  0.6792 -0.7497  0.1026  0.4231 -0.0262 -1.9647 -0.1665
#>  1.2325  0.2233  0.1197  0.7669 -0.4033  1.0722 -1.3553 -1.2556 -0.4756 -0.3690
#> -1.0225  0.6317  1.5627  0.3372 -0.6663  0.9605  1.2010 -0.4653  2.4829  2.3344
#> -1.0944  1.0722  0.1784  1.3020 -1.2178  1.4751 -0.8244  0.1055  0.1076 -0.3366
#> -0.2863 -0.2175 -0.7326  0.8970  0.7307  0.7016  0.7262  0.2753 -0.2105 -0.9918
#> -1.2686 -0.8631  1.7458 -0.0341 -2.0155  0.8349 -0.2215  0.3988 -1.3351 -1.0357
#> -0.1489 -0.5008  0.1681 -0.1252  2.0315 -0.8286 -1.5355 -2.0062 -0.5078  0.2134
#> -0.9401 -0.2821  0.5010 -1.1268 -2.2416 -0.8324  2.1799  0.7912 -0.3668  0.1398
#> -1.4772  1.3724  1.6403  1.0872  2.1507 -1.5835  1.7612 -1.5726 -0.0204  1.3859
#> -0.7120  0.7389 -2.0307  1.0143  1.3081 -0.2939 -0.5926  1.3494  0.5615 -0.1528
#>  2.4989 -0.1013 -0.3746 -0.0534 -0.3228 -0.1676 -1.0194  0.7676  0.9796 -0.5159
#>  0.9623 -2.0409  0.4067 -1.7821  0.5131 -0.9410 -0.2287  0.2979  0.6558 -1.2708
#>  1.9148  0.0691 -0.5941  1.4196  1.5129  1.1818  0.6475  0.8809 -1.1443 -1.5184
#>  1.2042 -1.3040  0.2893  0.0897  0.7056 -1.6486  1.0416  1.5018  0.6613 -1.5003
#>  0.9470 -0.4378 -1.3335 -0.3426 -0.9090  1.5609  0.8157 -1.0688  0.4953  0.6277
#>  0.6986  0.8805 -0.2568  0.1848  0.1448  1.1147 -0.7698 -0.6507 -0.7267 -0.3103
#> -3.4822  0.9246  1.2849  0.0181 -0.3807 -0.9840  0.6521 -0.3502  1.5821 -1.2085
#> -1.4281 -0.7964  0.1388 -0.8827  0.1905 -0.9089  0.3786  1.4662  2.2395  0.5609
#> -0.6584  0.2890  0.1154  0.4263  1.7707 -2.4033  0.0411  0.6204  1.4698 -1.3104
#> -1.0471 -0.3513 -0.7534 -0.6061 -0.6217 -0.0990 -0.2227 -0.6602 -0.9668 -1.0731
#> -1.6426 -0.9849 -2.4607  0.9203  0.8258 -0.4703  2.0387  1.5370 -1.1445  1.3484
#>  0.6217 -0.3262  1.7709  0.5522  0.5564  1.4094 -1.8829  0.2569  0.4349  0.8165
#> -0.3017  0.2678  0.5582  0.3440 -0.8608 -0.6479 -2.0541 -1.7538 -0.5587 -0.8291
#>  1.2377 -1.0450 -0.1064 -0.7915 -0.6917 -0.6334 -1.2728  0.8964  0.0855  1.1207
#> -1.0311 -1.9492 -1.3103 -0.1058  0.4341 -1.6672  0.7663 -2.1766 -0.4447  0.5396
#> -1.5246 -0.2572 -0.1817  0.5890 -0.6851  1.0401 -0.6158 -0.4386 -0.4296  0.5067
#>  1.1826  1.2552  1.5922 -1.1158 -0.0738  1.2792  1.5122  0.8190  0.1517  2.3375
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]