Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.9254 -0.6584 -0.9588  1.0306 -1.8555 -0.1389 -1.4531  0.4384  0.4472 -0.9385
#>  0.6805  0.0383  0.1926  0.1277 -0.4614  1.0074  1.7785 -0.3247 -0.7586 -0.9114
#>  0.4640 -0.0657  0.1025  1.1845  1.1874  0.6059  0.1845  0.2106 -0.5300  0.3545
#>  0.3969 -0.3722 -0.7001  0.7926  0.0765 -0.4099  0.7129 -1.1054  0.1891  1.4384
#>  2.2346  0.7525 -0.9493  0.7513 -0.0888 -0.3491  0.0557  0.7930  0.1116  0.7521
#>  1.3612 -1.3937  0.2312  0.3250  0.8097 -1.1561  1.4319  1.5472  0.8575 -0.3370
#>  0.7261  0.2593  1.3734 -0.8331 -1.5291 -0.7872  0.7927  0.6731 -1.5859  0.3386
#>  0.7228 -1.5416 -0.9238 -0.6406 -0.2714  1.4770  1.5728 -1.3929 -1.5078  0.2323
#> -1.9178  0.2790 -0.3395  1.0097  1.0510 -1.0160  0.2891  0.9920  1.5765  0.2375
#>  0.0659 -0.1266  0.9310 -0.2157 -0.7306  0.4035 -0.3926 -0.8479 -0.4581  0.7478
#> -0.3519 -0.0672 -0.7734  0.6278 -1.0465 -0.7873 -1.0758  1.3005  0.7246 -0.2396
#>  0.4241  0.1973 -1.2476  0.9887  0.5894  0.6341 -1.0594 -2.0198 -0.0117 -0.7338
#> -0.1214 -0.3375 -1.0768 -1.5384  0.2773  0.6226  0.1962 -0.5015 -0.0151  0.1398
#> -0.6505  1.2187 -0.5204  1.4768  0.1535 -1.2700  0.5326 -0.2332  1.6910 -0.7417
#>  1.3039  1.0453  1.4690  1.7204 -0.0706  1.3913  1.1446 -1.9438  1.4539  0.8885
#>  0.5241  0.5254 -1.7100  0.0959  1.3021 -0.3574 -1.2330 -0.1629 -1.2604  0.4286
#> -0.9920 -0.7750  0.5995  0.9996 -0.2284 -0.7191  0.5382  0.5574 -0.2911 -0.4299
#> -0.5785 -0.5589  0.1103  1.8106  1.8466 -0.4162  0.4567 -0.2794  0.0417  0.6321
#>  1.2930  1.2319  0.1154 -1.7139 -1.9155 -0.4307 -0.9124 -0.1216  1.4304 -0.5304
#> -0.1594 -0.0091 -0.0183 -1.6463 -0.9576  0.3772  1.1702  0.6447  0.1138  0.6645
#>  0.4974  0.2284  0.0622 -1.1091  1.8570 -1.0840  1.2863  2.3001  0.1685  0.7144
#>  0.5602 -1.1800  0.2734  0.9353 -0.6560  0.1286  0.8356 -0.1270 -0.2670 -0.0805
#>  1.2128  1.1017 -0.9520  1.0753 -0.4549  0.6532 -1.1833 -0.2058 -0.3586 -0.9937
#> -0.5493  0.1781 -1.3530  0.5372 -0.4778 -0.9502 -0.7431 -0.1394  0.5330  0.2489
#> -0.4408  0.1836 -0.7418 -0.1770  0.2896 -1.2346 -0.8049  1.3170 -0.0893 -0.4469
#> -0.7170  0.2669  1.1929 -0.1278 -0.1725  0.4230  0.2728 -0.5118 -0.4410 -0.0458
#> -0.5706 -0.1021 -0.4453  1.1262 -1.5903 -0.1947  1.4137 -1.3711 -1.3725 -1.1888
#> -0.5433 -0.8692  1.6055 -0.3276  0.6665 -0.6232 -1.5879 -1.0586  0.7967  0.0759
#> -0.1670  1.0380  1.5805 -0.5694  0.8879 -0.7931 -0.4090 -1.2620 -0.1935 -0.9887
#>  0.3348  0.3146 -1.0988  1.4935 -0.2074 -0.9594 -0.1670  0.2323 -0.8541  0.1256
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]