Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.6598 -1.9229 -2.4306  1.5440 -1.2263  0.0756 -0.1793  0.3619  0.2897 -1.1085
#>  2.0587  0.1044  0.1956 -1.1407  0.3634  1.3356  0.1434  0.0666  0.2501 -0.4951
#>  0.9334 -0.5359  1.5028 -0.8768 -0.0600  1.2893 -1.5750  1.3300  1.2995  0.0820
#>  1.7675 -1.1713 -0.3319  0.4458 -0.2253  0.6565  0.3523  0.2848  0.3851  0.4998
#> -0.3392  0.3847 -1.0974 -0.2774  0.0095 -2.0103 -1.3749 -0.0496  0.3807 -2.3656
#> -2.3065 -0.2051  0.8363  0.8093  0.1812 -0.0460  0.3357 -0.0640 -0.5257  1.5569
#> -0.4484  0.3755  0.5336  1.4308 -0.6372 -0.3866 -3.1239 -0.5119 -2.2007 -0.8132
#>  0.4470  0.3232 -0.2626  0.1975 -0.0883 -0.5082  0.0072 -0.2785 -2.7589  0.7100
#> -2.1719 -0.4411 -1.1010 -0.4607 -0.6230  0.9475 -0.1845 -0.4385 -0.2879  0.8828
#>  0.5904  0.0540 -0.1932  0.2217 -0.2688  0.0010  0.8687 -0.6596  0.5568 -0.8518
#>  1.5109 -0.9471 -0.2863 -0.5072  2.0246  0.0352 -1.1560 -0.2024  0.0842 -1.0103
#> -1.0059 -0.6242  0.1572  0.6710 -1.1323  1.3111  1.0767 -1.2223 -0.2898 -1.1272
#> -0.5441 -0.4603 -1.6167 -0.4916  1.3370  0.5818 -1.2303 -0.4235  0.0919 -1.0446
#>  0.4507 -1.1650 -0.4481  0.1245  0.3780  0.2675 -0.1283  0.9928 -1.1814 -0.1723
#>  0.2959 -0.3124  0.6270  0.6311 -0.1659  1.3774  0.4234 -0.9321 -0.0474  1.1616
#> -0.7083  0.3582  0.4884  0.2391 -0.3256  1.0813 -0.7313 -0.0615  0.0314 -0.4348
#>  1.5828 -0.2398  0.8894  1.0238  0.6099  0.6125  0.3369 -0.8614  0.5735  0.4391
#>  2.0219  0.6935  0.5515  0.3346 -0.3033  0.5954 -0.3633  0.5035 -1.5977  0.1539
#>  0.2395 -0.6162 -0.0305  0.9255  0.3263 -0.2368 -0.6826 -2.8237 -1.0882  0.0804
#> -0.5477 -1.5278 -0.4913  1.6890  0.2296  0.1962 -1.0251 -0.0884 -1.1576 -0.6519
#> -0.6935  0.2604 -1.4037 -0.2496  0.3647 -0.5287 -1.5606 -0.9044  1.0476 -1.4477
#>  0.2843  0.2514  0.3853  1.8343  0.3342 -1.1113  1.9584 -0.1371  1.3311  0.5322
#> -0.3994 -1.8636 -1.4941 -0.4846  0.2230  1.1021  0.6372 -1.5725 -0.2660  0.5472
#> -1.0074 -0.5394  0.4573 -1.2602 -1.7403 -1.3316 -0.9206 -0.9833 -0.1109  1.5222
#> -0.4897  0.0510  0.1674 -0.7094 -1.0678 -1.3221  0.2553  1.7541  0.1812 -1.6345
#>  1.8214 -0.6396 -2.3885  0.2257  1.4668  0.0234  1.0678 -1.3726 -0.0360 -0.9221
#> -0.2125  1.6511  0.4569  0.3601 -1.2267  1.0360  0.3445 -1.0363  0.8864  0.4322
#>  1.0920 -0.3265  1.2662 -0.2317  0.7485 -0.2043  1.7831 -1.3287 -1.0183  0.4614
#>  0.2946 -0.4923  0.9607  1.7322  1.9183 -0.0557  0.9571 -0.1652 -1.8423 -0.7778
#>  1.1442  2.1182  0.8848 -0.1366  0.9428 -0.6809  1.0344  0.1448  0.0156 -0.4866
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]