Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.2084 -2.0421  0.6130  0.7992 -0.1877  0.0255 -0.9354 -0.4485  0.4658 -1.2282
#> -0.6638  0.3870  0.6412 -1.4232  0.2611 -0.4819 -0.5327 -0.1613 -1.8139  0.5069
#>  0.6206  0.7732  0.2985  1.4748  0.3019  0.0292 -0.0853  0.6207 -1.6898  0.4999
#>  0.4551 -0.4633 -2.7485  0.5451 -0.3749 -0.8583 -1.0595 -0.6539  0.8917  0.9485
#>  0.8157  1.3639 -0.4858  1.0933 -0.1663 -0.7143  0.1598 -0.9490 -1.6341 -3.1340
#> -0.2106  0.7205 -0.8048  1.4793  0.0394 -0.5076  1.2259 -0.7397  0.0972 -0.2058
#>  1.1808  0.6611 -0.5591 -1.1134  1.6937 -0.4732 -0.2905  1.0510 -1.4952 -0.1345
#> -0.2015 -1.1781 -1.7299  1.0299 -0.7748 -0.8957 -0.3456 -0.3654  0.4821 -0.0716
#>  0.0972 -0.3277  0.2880 -2.0997  1.2292 -0.3547  0.7518 -0.0648  1.0592 -0.7560
#>  0.2767 -0.4424  0.6298 -0.8017  0.3189 -0.7234 -1.1187  1.3449 -0.3062 -1.6899
#> -1.6098  0.2487 -0.3568 -0.5130  0.8024 -1.1411 -0.2007  2.6332 -0.4685  1.8533
#>  1.1654  1.3604  0.2445 -1.6470  0.8777  0.6872  3.1541  0.7401  0.1510 -0.5225
#> -1.3931 -1.6808 -0.9439 -1.1314  0.6703  1.3462 -1.6405 -0.6130 -0.4021 -1.4678
#> -0.0078  0.7672 -2.4483  0.1145  1.1200  1.8185  1.6358 -0.6325  0.1270  1.2954
#> -0.7766 -0.4359  0.8555  0.4579 -0.4214 -1.2271 -0.2239  0.6338 -0.3542  0.2693
#> -0.0105  0.4624  0.1679  0.4119  0.3961 -1.0885 -1.8275 -0.0332  0.6739 -0.5904
#> -0.9173  0.4512  1.3803  0.8152 -0.8907  0.8282 -0.1728 -0.8973  0.7741  0.0195
#> -1.1765  1.3874 -0.5150 -2.4620  1.8203 -0.2273  1.4601  0.7274  0.5537 -0.0343
#>  1.3437  0.2765  2.1560 -0.0394 -0.2354 -0.0768  0.4590 -1.3386  1.0471 -0.4782
#> -1.6724 -0.5146  1.1941  0.6429  0.1817  0.7805  0.9702 -0.4175 -0.6588 -0.9915
#>  1.1392 -0.0056  0.2864 -1.0460  0.1131  0.7051 -1.3997  0.4124 -0.2205 -0.5084
#> -0.3194 -0.4045 -0.2964  0.5133  0.5589  1.5601  0.5133  0.6366  0.0714  0.1261
#> -1.7525  0.8836  1.2294 -1.1409 -0.1420 -0.6020  0.9622 -0.7306  1.6923  1.0149
#> -0.2818  1.0803 -0.8400 -0.3514  0.7489 -0.6970  0.3012  0.0638  2.5569 -1.3276
#>  0.5734  0.4932 -1.2484 -0.1454 -1.7887 -2.3284 -0.0097  0.7065  0.4308 -0.7988
#> -0.0818  1.5134  0.3105  1.3414  0.9011  0.3924  0.4211 -0.4773 -0.0043 -0.9943
#> -1.3668 -1.4238 -0.2497  0.4685 -0.6663  0.0808 -0.4107  0.6542 -0.4074 -0.9788
#> -0.6585  0.4215  0.4366  0.8894  0.4046  0.0542  1.1941 -1.5775 -0.3258  0.3901
#> -1.2758  0.5040  1.1955 -0.2975 -0.1810  0.3212 -0.5245 -0.4923 -0.4271  0.3241
#> -0.6546  0.7751 -0.8063  1.4860  0.8635 -1.2000  0.4819  0.3881  0.7682  2.5179
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]