Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2389  0.4877 -1.2855 -0.9789  0.9716 -1.0862  0.2996 -0.0898 -2.2800  0.1861
#> -1.0463  1.2610 -0.5135  0.0341 -0.3666  0.1013 -0.6592 -0.4211  0.0984  0.9194
#>  1.1813 -0.7946  2.3062  1.3341  0.9809 -1.3842 -0.2252  2.2308 -1.2067 -1.8254
#> -0.2275 -0.8663  0.1099 -1.5026  0.1401 -0.2269 -1.0788  1.0117 -1.0996  0.8797
#>  0.5554  1.4841 -1.5234  2.3393 -0.6635  0.7166  0.3764 -0.7655 -0.7115 -0.5357
#> -1.5405 -0.9161 -0.6744  0.2851  0.1023 -0.8880  0.3021 -0.2934 -2.0137  0.1517
#>  1.0482 -1.1232  1.5499  0.3181  0.0671  0.6416  0.4822  1.2385  0.0672  0.9112
#>  0.8132 -0.0696  0.0557 -0.4639  0.1544 -0.2791  1.6778  0.6635  0.2690 -1.1287
#> -0.8015 -0.3936  0.1123  1.5704  0.5736 -1.9327  0.7359  1.2986  0.4588  2.3254
#>  1.3654 -0.1454 -0.0567  2.3015 -0.3068 -0.3294 -0.1688 -1.7441  0.1892  0.8046
#>  0.6739  1.4603  1.2322 -1.5484  1.1081 -0.4717  0.9520  0.4114  0.3762  2.6402
#>  0.7818  1.6203  1.0619 -1.1139 -1.4885 -1.0679  0.3034  0.1961 -0.2603 -1.8916
#> -0.0968  1.0520  0.4043 -0.4042 -1.5273 -1.2939 -0.0290  1.5035  0.1741  1.3108
#>  0.2592 -2.4120 -0.4969 -1.0551  0.3658 -0.0138  0.1883  0.5189  1.5222  1.1199
#>  1.0564 -1.0351 -0.3162  0.4313 -0.1502 -1.1665  0.1585 -0.1056  0.8545 -0.4880
#>  0.4630  0.6628 -0.6105  1.0856  1.3865  0.6892  1.2902 -1.0313 -0.4829 -1.8821
#> -1.0595 -1.0154  0.6825  1.3407 -0.0934 -0.8704  0.2839  1.8655  0.7761  0.6412
#> -0.0190 -0.8094  0.8081  1.7393  0.8648 -0.5984  0.1088  0.5654  0.5028  2.2870
#>  1.1028  1.4309  0.2025  1.3509  0.2430 -0.4776  1.0913 -0.2105 -0.3852 -0.4898
#>  0.7066  0.8174 -0.1059  0.1993  0.1407  2.3762  0.7323  1.8977 -0.6448 -0.9970
#> -0.7174  0.6400  1.6141  0.4039  0.4440 -1.2783  1.3237  1.5120  0.8982  0.4564
#> -0.1478  0.4269  1.6801  0.0686 -0.1124  1.0055  0.3225  0.5956 -0.8871  0.7848
#> -0.5059 -0.5324 -0.0346  1.1883 -0.0246 -1.7092  0.5719 -2.0945  1.1188  0.0061
#> -1.4302 -1.1705  0.9376 -0.3107 -1.0842 -0.3136 -0.8050  0.3419 -0.6663  1.0587
#> -1.1400  0.6749  0.8571  0.9373 -0.2283  1.8047  0.9629 -0.6648 -0.6987 -0.1792
#> -0.2806  0.6273  1.3862  1.7579 -0.6134 -1.2394  0.1128  0.7138 -0.5561  1.0067
#>  0.2846 -1.0662 -0.7537  0.1890 -1.4315  0.7185  1.2367 -0.0464  0.2006 -0.7190
#>  0.8928 -0.6621  1.0762 -0.6488  0.3970  1.0860  0.3749  0.4791  2.1675 -1.1779
#> -0.1082 -0.3746  0.8655 -0.2729  1.2119 -0.8798  1.1671 -0.5125  0.9242 -0.7824
#> -1.4716  0.7171  0.3357  2.1648  0.3336  1.7215 -0.3428 -0.2927 -0.6184 -0.7458
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]