Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2693  0.4522 -0.1618  0.6803 -0.6322  0.0011  0.4282 -0.9567  0.0359  1.2386
#> -0.0722 -0.5189 -0.9238  0.3075 -0.8459 -0.6541  2.2648  0.2597 -0.2586 -1.6694
#> -1.1316  0.4239  2.5741 -0.3389 -0.5687 -0.0765  0.4886 -0.5119 -0.9086 -0.5055
#>  0.1666  0.2157 -0.2611 -0.8891  0.3433 -0.2550 -0.9370  0.3732 -1.6415 -0.3632
#>  0.6370  1.2476 -0.6118 -1.9139 -0.0024  0.4236  1.5867 -0.9258 -1.0677  0.3389
#>  0.5347 -0.8926 -1.4360  0.2124 -1.2016  1.6460 -0.7931  0.4248 -1.6782  0.0962
#> -0.7209 -0.5769 -1.8081  0.2703  0.2674 -1.1592  0.1981  0.4947 -0.1804  0.6707
#>  1.1327  0.0107  0.3390 -1.3726  2.0623 -0.7900  0.0457  0.7325  1.0602  0.5709
#> -1.5264  0.7485  1.0633  0.0742 -1.6856 -1.7152  0.9531  1.2441  0.4835  1.0317
#>  0.3174 -0.3799 -1.5142  0.5894 -0.4486  0.6949 -0.5186  0.4635  0.9423  0.2832
#> -0.6898 -2.3809  2.0634 -1.2559  1.0767 -0.8154  0.1903 -0.0973  0.7145 -1.3324
#> -1.4097 -0.0480 -0.1309  0.1457  0.7907 -0.7931 -0.9277  0.7853 -0.0726  1.0686
#> -1.7804  0.9810 -0.7510 -0.8464 -0.0082 -0.7796 -0.8650 -0.9824 -0.7062  0.8415
#>  0.3946  0.9327 -3.1750  0.1043  0.8640 -1.7523 -0.9138 -0.6627  0.0345  0.4594
#> -1.0151 -0.2539 -0.3829 -0.4229 -0.0470 -0.8101 -1.1719 -0.8768  0.9460  0.2903
#> -1.2338  0.1607  0.8984 -2.0959  0.5785 -0.2390  1.0267  0.7861 -0.8115 -1.5866
#> -1.6743  0.3907  1.7139 -0.2729 -1.8744  0.1409 -0.7373  0.9126 -0.1253 -0.0743
#> -2.2908  0.9392 -0.8176  1.2580 -1.0705  1.0984  0.5013  1.4383  0.0636 -0.6724
#>  0.4124 -0.8517  0.6126 -0.2056  0.6437  0.1522  2.0166  0.0829 -1.2100  1.3732
#> -1.9943  0.9698 -1.2764  0.7353 -1.6045  0.0802  0.7437  0.1825 -0.9530  2.0318
#>  0.5487  0.8767 -1.3057 -0.2049  1.1944  1.7768 -1.4010 -0.6391  0.6259 -0.1936
#> -0.9548  0.4453 -0.0937 -0.5740  0.4939 -1.2058  0.1073 -0.3414 -0.6255 -0.0541
#> -1.3713 -0.7111 -0.3153  0.6419  1.3813 -1.0814 -0.4450  0.6655 -0.1320  0.4688
#> -0.0281  0.7965 -0.6289 -0.5865 -0.2276  0.4985  0.1784 -0.9815 -0.8470 -0.5559
#>  1.2595 -1.0466 -2.3007  1.5395  2.0289  1.5862 -0.7370 -1.2612  0.8975 -1.2095
#>  1.8997  0.8312 -0.8159 -2.4139  0.4085 -1.4833  1.2862  0.9570  1.6358  2.2562
#> -1.2455  1.0015  0.5993  0.8501 -0.8003  2.5073  0.1984  1.7717 -1.3954  0.3904
#> -0.0859  0.6630  0.0742  0.4608 -0.1954  1.0915 -1.6151 -1.7798  0.8418 -1.1292
#> -0.8793 -0.3779  0.4120 -0.4210  0.5495  0.7960  0.7526  0.9895 -0.6144  0.7939
#> -0.0410  1.2532  1.0450  0.2760  1.4457 -0.6260  0.7723  1.0748 -0.2542 -0.1907
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]