Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.4725 -1.8417 -0.1175  1.2329 -1.0247  0.3918  0.1019  1.4198  0.4642  0.8516
#>  0.4643 -0.5549 -0.2357  0.8390  0.3515 -0.2805  0.1064  0.7480  1.9418  0.7178
#> -0.1234  0.7147 -1.0091 -0.8639  1.8767 -1.4523 -0.9059  0.7338 -1.2623 -1.1465
#> -0.1339 -1.0882 -0.6061  1.6324 -0.6369  2.4361  1.1386 -0.1293 -0.2883  0.4713
#> -0.2521 -0.1399  0.4203  0.6440  1.2577 -0.1487 -0.7023  0.5025  0.6818  0.4334
#>  0.3225 -0.0812 -1.7094 -1.2215  0.0473 -0.2907 -0.6643 -0.1774 -0.7464  0.2449
#> -0.6958 -0.0602 -0.9985  0.1053  0.5195  0.6689  0.9538 -0.7915 -0.0731  0.7522
#>  1.3067 -0.1344 -0.9567 -0.7854 -0.8449  0.0030  1.5296  0.6629 -0.8280 -1.2444
#>  0.4041  0.6366  0.3372  1.4186  0.3024  0.0174 -2.0808 -0.7100  1.9175 -0.5984
#>  1.4556 -0.8718 -1.0023 -1.1068  0.6621 -1.0669 -1.7146  2.0015  0.4328 -1.6012
#> -0.9509 -0.5164  1.0288  0.5333  0.4296 -1.0848  0.1323  0.1432 -0.6994  0.4589
#> -0.8386  0.9850 -1.2808  1.2540  1.1168 -0.6393  1.1528 -0.4074 -1.2327 -0.4835
#> -0.4268  0.3532  0.3268  0.2897  0.0737  0.9467 -1.7554 -1.4032 -0.8392  0.4109
#>  0.8044  1.0728  1.1106  0.1512  0.6761 -0.4343  0.2015  1.8492  1.7770 -1.8358
#>  2.1040 -1.2731  0.8510 -0.5080 -2.7861 -0.2524  0.7837 -1.1366 -1.3497 -0.2293
#> -0.9495  0.0215  0.6650 -0.4300 -0.3651  0.1726 -0.5367  0.2059  0.2428 -0.3981
#>  0.4932  0.3267 -0.1768  0.2117 -0.3484 -0.0455 -0.3027 -0.3234 -0.0041  0.2257
#> -0.4651 -1.4296  1.6174 -0.9231 -1.1875 -1.6567  1.2227  0.2614  1.1802 -1.2317
#>  0.2300 -1.5751 -0.1179  0.8555 -0.0273  0.0267 -1.3528 -1.5171 -2.0398 -0.0440
#>  0.1511 -0.5730  0.1541 -1.2736 -0.5138  0.0095 -1.6596  0.2078  0.5214  2.3899
#>  1.7425 -1.0492 -1.1580 -0.3249  1.1760 -1.4522 -0.4951 -1.5632 -0.8004  2.5729
#>  0.2540  0.9982  0.6352 -0.6986  0.4859 -0.0393 -1.7973 -0.5095 -0.2102  0.5277
#>  0.8704  1.2102  0.6509  0.1294  1.1361 -1.3314 -0.9632  0.1868  0.8255 -1.2693
#>  1.0353  1.4377  0.3861  0.6213 -0.7727 -0.1286  0.5285 -0.9678  0.7543  0.1311
#>  1.2669  0.8114 -2.2145  0.9952 -0.3889 -0.0428 -0.2705  0.1387  0.1704  0.9501
#>  0.8914  0.4238 -0.1782 -0.5490 -0.9712  0.6566  0.3584 -2.6501  2.1861  0.0719
#>  0.9820 -0.2345  0.9776 -0.6755  0.5713  0.8830 -0.6371 -0.3131  0.8286  0.3510
#>  0.8048 -1.8240  0.6204  1.5447 -1.3904 -1.2038 -0.9867 -0.4112  0.8453  0.9741
#> -0.2377 -0.7147  0.5192 -1.9372 -1.9020  2.1116 -1.2110 -0.7560  0.1495  0.3286
#> -0.4201  1.2153  0.5176 -0.5425 -0.4443  1.3235  0.5814  0.1120  0.3550 -0.2514
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]