Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.9228  0.4121  0.3520 -0.9817  0.5987 -0.8106 -1.3819  1.1840  0.2975  0.2156
#> -1.0409  0.2192 -0.9117  0.5804  0.3530 -0.5737  0.6156 -0.7713 -0.7383 -0.6502
#> -0.8957  1.5699 -1.0091  0.8609  1.2592  1.3364  0.1049  0.0161  0.3557  0.5453
#> -0.1363 -2.2550  0.6002  0.4192 -1.5737 -0.3662  0.0235  0.3415  0.7307 -1.1802
#>  1.7766 -1.0247 -1.5090  0.1055  0.9679 -0.8295 -0.4293 -0.1969  1.4940  0.0704
#> -1.2536 -1.8076 -0.4657 -0.1027 -0.5718 -0.0066 -0.9369 -1.0205 -0.1276 -1.3733
#>  2.1884  0.2236  0.1832  1.2581 -0.7318 -0.3854 -2.0585  0.3144  0.7164 -1.0938
#> -1.3419  0.7123  0.2854  0.1839  0.5237  0.5162 -0.9070 -0.4125  0.8228 -0.4473
#>  0.9774 -0.4471  0.5436 -0.0493  0.4873  0.2980 -0.3157 -2.6770  1.0040  0.9454
#> -0.0338 -0.7599  1.3246  0.1336  0.1522 -0.2003 -0.9990  1.7065 -0.1962  0.0761
#> -0.0754 -0.7401 -0.3083  0.8651 -2.0411 -0.2402 -0.2689 -2.0537  0.2227  1.1460
#>  1.0773  1.0294  1.0323  0.1013  0.3075  0.1048  0.5394  0.5544 -1.7327  1.8703
#> -0.6143  1.3470  0.0875  0.7633 -0.8793 -0.3233  1.5643 -1.8180 -1.0890  0.6397
#>  0.3326  0.8815 -0.8223 -0.5246 -0.2576  0.2958 -1.4193 -0.6866 -1.8668 -0.6137
#>  1.1879 -0.5611  1.1225  0.5180  2.2660  2.6665 -0.8534  0.7031 -1.2165 -0.0188
#> -0.2180 -1.2616  1.0840  0.4691  0.2706  1.0808  0.7980  0.9091  0.0533  0.5884
#>  0.9940 -1.7401  1.3514 -1.3610  0.7431  0.8291  0.4957  0.1306 -0.9947  0.9888
#> -1.2313  0.8535  0.0376 -0.8398 -1.3598 -0.5511  0.4954 -0.6598 -0.3662 -0.4493
#>  1.6187  0.4389 -0.5712 -0.9793 -0.1071  0.3678 -0.3968 -2.0634 -0.3412 -0.1788
#> -0.6305 -0.7717  0.3218 -0.9618  2.1795  0.3545  1.6030 -2.1966  1.7619 -0.2532
#> -0.4958 -1.4716  0.0126  0.0923  0.4399 -1.2174  0.9892  0.5774 -0.6086  1.8495
#> -0.3069  0.7771 -1.3379 -1.7853 -0.5410  0.4718 -0.4429 -0.1249 -0.5201 -1.9236
#> -0.8683  0.5881 -0.3442  0.1479 -1.0886 -1.3854 -0.7789 -1.6646 -0.9573 -1.2681
#>  0.2433 -1.1090  0.7964  0.0937  0.4543 -0.1675 -0.8067 -0.0878 -0.0037 -0.9501
#>  0.2284 -1.7010 -0.8463 -1.9364  0.4283  0.8306 -1.5153  1.0169  0.3769 -0.2786
#> -1.3019  0.8999  0.0532  1.6348 -1.3358 -0.3940 -1.6986 -1.3373  0.6651  1.3955
#> -1.7216 -0.2900 -1.2278 -2.1271  0.3878 -0.9556 -1.5911  0.5116  0.0710  1.8183
#>  0.7429  0.0760  0.9648  0.3205  0.1979  1.2548  0.3451  1.3917  0.5461  0.1996
#>  0.8224  0.8245 -0.0363 -0.9866 -0.7546 -0.1879  0.5730 -0.9876  1.8824 -0.4339
#>  1.5206 -0.5987  1.8964 -0.7505 -1.5062  1.1640  1.3412 -1.6087 -1.0164 -0.3743
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]