Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0845  0.5538 -0.4526 -0.3792  0.2618 -0.2003  1.2516 -0.7315  1.2289 -0.2744
#>  0.8897 -1.0920 -1.4419 -0.4006 -0.9720  0.1694 -0.8871 -0.9352 -1.7234  1.2965
#> -0.8160  0.2967  0.3306  0.4435  0.0692  0.3693 -2.4250 -1.2321 -0.0354  1.2102
#>  0.2517  0.9677  2.3033 -1.1475 -0.9972  0.8170 -0.5447 -0.7509 -0.0290 -0.9616
#>  0.0961 -0.2409 -0.5922 -1.0665 -1.0007  0.9366 -1.3140 -0.3794  0.6601 -2.2771
#>  0.5809 -0.1473  0.7671  0.1641 -0.2090 -0.4856 -0.6746 -1.0269  1.6285 -0.7439
#>  1.5017  0.3696  1.2545 -0.3398  0.0901  1.8278 -0.4004 -0.2834  0.3057  0.8444
#> -0.9699  1.3773  0.9006 -0.5354  0.5229  0.7879  0.0481  1.0634 -0.2978 -0.8241
#>  0.0955 -0.1036  0.9084 -0.1639 -0.0071 -0.5148  0.3056 -0.8584  0.2492 -0.6851
#>  0.3210  2.2591 -0.9931 -0.7324  0.2102 -0.7135  0.4468  0.2137  0.9364 -1.3712
#> -0.0400  1.5455  1.7785  1.1272 -2.1781 -1.7651  0.6597  0.5541 -0.3746 -1.0905
#> -1.2728  0.3602  0.5291  1.1611 -0.3191  0.2542 -1.2587 -0.3744  0.3996 -1.6517
#> -0.0973 -0.6197 -0.0141 -0.9546 -0.3403  1.1543  0.9520  0.9566  2.8500 -0.1880
#> -0.0594 -0.4188 -0.2403  1.0520  1.5195  2.3969 -0.4344  1.0529  0.1558  1.6316
#> -1.5231  0.4950  0.7310  0.4771 -0.4416 -1.3371  1.5568  0.2432 -0.6605  0.0956
#> -1.0597 -0.6042  1.3945  1.6845  0.3137  1.6436  0.5243  0.5172  0.4349  0.5408
#> -0.0318 -0.8018 -1.1608 -1.3554 -0.8186  0.4151  1.3330  1.6115  0.5135 -0.5325
#> -0.4198  1.3754  0.8167 -0.2345 -1.8160  0.3360  1.4890  0.2449  0.1307 -0.5049
#> -1.7237  0.2649  0.6300  0.0460  1.2855  0.9916 -0.5802  0.4437  0.2163 -1.0765
#> -0.9707  0.9369 -1.1645 -1.1981 -1.5230  0.3735 -0.1382 -0.5147 -0.8081 -0.4480
#> -0.2200 -1.3469 -1.0687  0.0731  0.8680  0.6498 -0.7476  0.1772  0.1704  0.9552
#>  0.1534  0.4889  1.1327 -0.5933 -0.7615  0.8365 -0.2267  0.3784  0.1365  1.4250
#>  0.6346 -0.0189  0.1519 -0.7724 -0.9304  0.2564 -0.9684  1.3955  0.0236 -0.6721
#> -1.3067 -1.4710 -0.6348 -0.1213  0.9081 -0.2835 -0.1677  0.4269 -0.5661 -1.0652
#>  0.5741 -0.5748  0.2277  0.6438  0.7314  1.8196 -1.0468  1.0515 -1.5711 -1.6659
#> -0.3838 -0.8712  0.6514 -1.0243 -0.6344 -0.0577 -0.7510 -0.4234  0.5252  0.5436
#>  1.7928 -1.0010 -1.0808  0.1817 -0.7232  0.0531 -0.8559 -0.0858 -0.4499  0.1134
#>  1.1482 -0.1135  0.1099 -1.4289 -0.4219 -0.5750 -0.1612 -0.0769  0.4486  0.3693
#>  0.0784 -0.7862  1.3299 -0.4362  0.0581 -0.0640 -0.8631 -1.1236  2.4940  0.7085
#> -0.0279  0.3349 -1.2006 -0.7102 -0.8442  0.2678  0.3511  1.5019 -0.3309  1.1609
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]