Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0968 -0.3674 -0.8512  1.9947 -0.7732 -0.6174 -0.0295  0.9361 -0.2238 -0.3831
#> -1.7505 -1.1474 -1.1224  1.4488  0.2645 -1.6161 -0.0290 -1.3438  0.3925  0.0616
#>  1.2432  0.2038 -0.7350  0.9102  1.0813  1.1217  0.5407  0.4859  1.0458 -0.1208
#> -1.4242 -0.0489  1.2628  0.5812 -1.1286  0.3265  1.0783  0.6514 -1.2480  1.0309
#>  0.4492 -2.6557  1.8827  0.5867  0.2492 -0.0808 -0.0262  2.1668 -0.0387 -0.7227
#> -0.6947  1.4009  0.4697  0.2110 -0.1148 -0.5510 -0.6039 -0.6572  1.0400 -0.9508
#> -0.2381 -0.8626  0.2676  2.6442 -0.1145  0.0957  1.7625 -1.9662 -1.1510 -0.9022
#> -1.8155 -1.6931 -0.1685 -0.4395  0.1289 -2.1653  0.1193 -1.5036  0.6127  1.6693
#>  0.1329 -0.0281  0.7637 -0.1180  1.0311  2.4692 -1.2959  0.4096  1.1733 -0.9801
#>  0.4875  1.2755 -0.5312 -2.0420  0.8911 -1.3241 -0.3775 -1.0026  1.1704  1.4062
#> -0.5727  1.7099  0.9752 -2.0746 -1.3256 -0.2260 -1.3283 -0.0243  0.2834  0.3030
#>  1.8836  1.2131 -0.4747 -1.3509  1.7804 -0.6693 -0.5803  1.1563  0.9318  0.6464
#>  0.6716  1.2533  1.8164  0.7041 -0.8518  1.2585 -0.2904 -0.6493 -0.6104 -0.6922
#> -0.9309  0.8110 -1.5919 -1.4539 -0.5524  1.3847  0.1801  0.7169 -0.6189 -0.2081
#> -1.6683 -0.4837 -0.2753 -0.5034  1.1995  0.0823  0.1508  0.6727  0.1315  0.6989
#>  2.5567  0.3177 -0.9737 -0.7429 -1.8513 -0.8765 -1.2888 -1.9131 -1.3229  0.5198
#> -1.3840  0.4762  1.3561 -0.3463 -0.2263 -1.4657  1.0782 -0.5121  0.1038 -0.5887
#>  0.1393  1.1230 -1.3070  1.2880  0.8450 -0.7840 -0.4663  2.2253  0.9255  1.7548
#>  0.1751  1.4623  0.0108 -1.6394 -0.2135 -1.7524  1.3766 -0.2153  0.7836  0.3360
#> -0.4106  0.8427 -0.7938  0.8605 -0.3486  0.0554 -0.5410 -1.0782  0.4752  1.0434
#>  1.0177 -0.1800 -1.0344  0.5840  0.5895  0.1834  0.1142 -0.9861  1.0256 -1.0420
#> -1.3286 -1.8534 -2.4981  2.5202 -0.9939 -0.4864  0.4011 -0.6970 -0.3995  1.9154
#>  0.5981 -0.9192  0.4840 -0.6080  0.6110  0.4867 -1.5757  0.4681  0.1074 -0.4211
#>  0.1289 -1.2509 -0.3433  0.0547 -0.0193  0.7849 -1.2234  1.0988  0.5848 -0.7770
#> -1.1319 -0.6081  0.0921  0.5378 -0.5808 -0.0561  1.6243 -1.0768  0.8093 -0.1012
#> -0.2108  0.5644  0.2926 -0.7380 -0.1967 -1.0542 -0.2879 -0.4295 -0.2496  1.0778
#>  0.3749 -0.3516  0.5615 -1.2878  0.3850  0.9798 -0.4243 -0.1782  0.2284  0.5737
#> -0.3681 -1.6899 -1.0872 -1.4315 -0.2430 -0.6260 -1.0184 -0.0463 -1.0629  0.8153
#> -0.3407  0.2455  0.4491  0.7925 -0.3098 -0.5054 -0.1622  1.6836 -0.4285  1.1757
#>  0.0802  0.5323  1.1781  0.3057  0.2866 -0.5594  0.3801 -0.6241  1.2981 -0.3241
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]