Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.4588  0.5857  1.2630 -0.0585 -0.5197  2.1999 -0.5435  0.5220  1.4438  0.3545
#> -2.2089 -0.1106  1.1256  0.2768  0.8137  0.2164 -0.6247 -3.7306 -0.4504 -0.0123
#>  0.6511  0.6887  0.2099  0.3856 -1.5173  0.1058  1.0466 -0.0556  1.9391 -0.4461
#>  1.1684 -1.7604 -1.1194  0.8112  0.0149 -0.2222  1.9233 -2.1372  0.0293 -0.1025
#>  1.0134 -0.1418  2.1775  0.5986  0.4786 -1.7516 -0.4983 -0.9059 -0.2449 -0.6757
#> -0.6663 -0.3048 -1.1045 -1.0808 -1.0420  1.1151  1.2706  0.0588  0.9515  2.1099
#>  0.5598  0.3845 -0.0456 -0.8800 -0.3982  0.1215  0.5159  1.1092  1.2587 -0.3568
#>  1.2495 -0.2324  1.2598  2.1837  1.5457  1.0762  1.9749  1.0681  1.5397  0.7547
#> -1.3594 -0.6498 -0.3249  0.4417 -1.9827  0.6672 -0.5495  0.8841  0.5672 -0.6091
#>  1.4295 -0.5856  2.7336 -0.0405  0.1097 -0.1350 -1.8465 -1.5899 -0.5220  1.5735
#>  0.9588 -1.2046 -0.1229  0.0188  0.0716 -0.0563 -0.4699  0.1385  0.4769  0.2565
#> -0.1873 -0.6639  0.9267 -0.0511 -1.2433 -0.1310  0.1827 -0.4713  0.4014  1.3107
#>  1.9214 -0.2084 -1.3098  0.3338 -1.2632 -0.2179  1.0317 -0.8551 -0.7393  1.3802
#>  0.1135  0.2599  0.5106 -0.9611 -0.9517  0.9912  0.1078  2.4195 -0.6536  0.5193
#> -0.8251 -0.4798 -1.0156 -0.4745 -0.2625  1.7179 -1.4566  0.0432 -1.2332 -0.0574
#>  2.0139  1.4094 -1.1072 -0.0894 -0.9237  0.0620 -1.0670  0.6029  0.0980 -1.3414
#>  1.0785 -0.5623  2.2582 -0.1241  0.2633  0.0498  0.5023 -0.2065 -0.0305  0.7256
#> -0.3683  0.0312  0.4120 -0.3705 -0.0219  0.4193  1.1407  0.7880  0.2859  1.0681
#> -0.3965 -0.0760 -2.2840 -0.0248  1.1474 -0.1976  2.2932  0.8865  2.0206  1.2378
#>  1.6634 -0.4979 -1.2161 -0.6201 -0.7311  0.6139 -1.7231  0.0206  0.0307 -0.3558
#>  0.7521 -0.3625  0.5842 -0.9215 -0.0741 -0.0840 -0.6966 -1.0042 -0.0549  1.7062
#>  2.7365 -0.4765 -0.4186  0.4978 -1.1766  0.0383  0.0829 -0.2213  0.7534 -0.3693
#>  0.4255  0.9659  1.1918  2.0648  0.9514  0.4083 -0.0025 -0.9885 -0.2343  0.0581
#> -0.4639 -1.0028  0.3307 -2.1125 -0.1023 -0.3498  1.7322 -0.1321  0.9402 -0.4441
#> -1.9121  1.1569  0.4365  0.0896  0.6321  0.8860  0.2939 -1.3841  1.2542 -0.2550
#>  0.1948 -0.6948 -0.1640  1.2651  0.0702  0.4160  0.7531  0.0074  0.5533  0.0874
#> -0.8509 -0.9883  0.1553  0.0477  0.4423  0.5784 -0.5109 -1.0036 -0.7483  0.2642
#> -0.8258  0.1766 -0.9591 -1.1725 -1.7736  1.5512 -0.4814 -1.4792  0.6401 -0.3740
#> -0.6216 -0.2180 -0.3079 -0.9646  0.3185 -1.2819 -0.6374 -0.2611  2.7946  1.6619
#> -0.2096 -0.3452 -0.9637 -0.2964 -1.1505 -1.1541 -0.6054  0.3165 -0.2411 -0.2449
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]