Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 6.5935e-01 -8.6541e-01 -1.0216e+00 -1.2916e+00  1.1255e+00  8.1364e-01
#> -5.9520e-01  9.0507e-03  4.3192e-02 -7.8939e-01  1.3561e-01  8.4956e-01
#>  1.0483e+00 -2.7404e+00 -7.6374e-01  1.8134e+00 -6.5462e-01  3.8797e-01
#>  3.0686e-01  1.0700e+00 -6.1226e-01 -7.6822e-02  7.0252e-01  1.0186e+00
#>  1.0061e+00 -1.0589e+00  7.5315e-01  9.7272e-03  6.5003e-01 -2.2413e+00
#> -1.2032e+00 -1.1982e+00 -5.3960e-01  1.1965e+00 -7.8238e-01  1.7223e-01
#>  2.2055e-01 -9.4929e-02 -4.0833e-01 -6.4356e-02 -1.4819e+00 -7.0263e-01
#> -2.8661e-03 -2.6248e-01  1.0786e+00 -6.5655e-03 -8.4196e-01  1.4072e+00
#> -6.6573e-01  1.5829e+00  9.0815e-01 -1.6278e+00 -1.0201e+00 -3.7876e-01
#> -1.1017e-02  4.7157e-01 -4.8441e-01 -5.9720e-01 -2.7220e-01  2.8042e-02
#> -2.3125e+00 -2.4344e+00 -6.3234e-01 -3.2016e-02  3.2367e+00  8.2722e-01
#>  4.7997e-01 -7.5852e-01 -1.3920e+00 -1.6173e+00  3.6074e-02  9.1597e-01
#> -2.3566e+00 -9.6276e-02  1.3924e-01 -7.9122e-01  1.8311e+00  3.1209e-03
#> -5.4008e-02 -1.8044e+00  3.4904e-01  1.0819e+00 -2.0725e+00 -6.2252e-01
#>  5.7098e-01 -2.8368e-01 -6.1587e-01  6.0479e-02 -3.0898e-01  5.4062e-01
#>  1.3061e+00 -5.3391e-01  1.1541e+00  1.3524e-01  1.0361e+00 -1.0418e+00
#> -1.1306e+00 -1.5056e+00 -1.6501e-01 -3.9423e-01 -3.2944e-01 -8.9902e-01
#> -2.2233e+00 -2.8549e-02 -3.1958e-01 -1.3068e+00 -1.1336e+00  1.0261e+00
#>  9.7294e-02 -7.4200e-01  9.0118e-01  1.0346e+00  1.9967e-01  4.6695e-02
#>  1.3666e-01  1.0250e+00  1.2935e+00  1.4804e+00 -2.6022e+00 -2.4213e-01
#>  6.6753e-01 -3.0833e-01  5.2631e-01 -1.0618e+00 -4.5275e-02 -1.1924e+00
#>  7.1154e-01 -1.8112e+00  9.5559e-01 -7.6723e-01 -3.3162e-02  1.0555e+00
#>  1.2891e-01 -6.0494e-01  8.4007e-01  1.8220e-02 -3.2699e-01 -2.1230e+00
#> -4.5126e-01 -1.0893e-01  8.0646e-01  2.0410e-01  4.7143e-01 -7.4713e-01
#>  4.7026e-01 -2.9537e-01  1.0807e+00 -6.4108e-01  5.7860e-01  8.7962e-01
#> -2.7478e-02  9.4902e-02  1.2617e+00 -3.7451e-01 -6.0533e-01 -1.7107e-01
#> -7.2673e-01  5.6867e-01  5.2216e-01 -2.1144e+00 -1.8334e+00 -2.3304e-01
#> -1.8810e+00  4.4656e-01  7.0358e-01  5.3409e-01 -7.2507e-01 -4.0222e-01
#> -1.9606e+00 -1.1771e+00 -9.8352e-01  4.0935e-01  6.9237e-01 -9.1991e-01
#>  9.0080e-01  5.2361e-01  1.0454e+00 -5.9296e-01 -7.2333e-01 -6.7190e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]