Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.6348 -0.0336  0.9867  1.1405  0.2668 -2.0970 -0.4295  0.5243  0.5959 -2.1018
#>  0.0076 -0.7961  0.7512  0.9540 -0.5517 -1.2573  2.1297  0.5108  0.2835 -0.8595
#>  1.5515 -0.3220  1.8497  2.0133 -1.0896  0.4703  2.6978  2.2435 -0.1071 -0.8271
#>  0.2264  0.5743  1.7067  1.1205 -0.4558  0.2056  1.0939 -0.4261  0.9520  0.8655
#>  0.1321  1.2938  0.6689 -1.2980 -0.7899  0.7368 -0.5270 -0.0277  0.6234 -1.2756
#> -0.6657 -0.3913  0.5538  0.4711 -0.1054 -2.0869  1.8319  1.1450 -1.4254  0.2570
#> -0.6548 -0.1910 -0.1812  0.0202  0.5792 -0.7684 -1.5277  1.7040 -0.8017  0.0710
#>  0.1020 -0.8465  0.9907  1.0723 -0.5519 -2.1642 -0.5658 -0.1397 -1.9794 -0.0357
#>  0.9047  0.3902 -2.3902  1.4600 -1.8071 -0.0052 -0.2357 -1.7364 -1.0647  0.5771
#> -0.3076 -0.8437  0.0663 -0.5819  0.1844 -0.6708  0.4077  1.4963 -1.1434 -1.5573
#> -0.3897 -1.6937  0.8458 -0.3166 -1.2084  0.1458  0.7035 -0.8739  2.1079 -0.1598
#>  0.9340 -0.0100  1.9757  0.1282  2.0362 -1.7383  0.3325 -0.1083 -0.3781  0.8117
#>  0.2240 -1.7536  0.0367  0.0519  0.2488 -0.5021 -0.1868 -1.5341  0.3952  0.4675
#> -2.2694  1.2642  0.1422  1.7939 -0.9147  0.8335 -1.3795 -0.6717  0.7019  0.8499
#> -0.9503  1.1869 -0.7255  0.2870 -0.3334  0.3738 -1.0932 -1.3055  1.0179 -0.1818
#>  0.4128 -0.3730 -0.1621 -0.2829  1.4883 -0.6908 -0.6325 -0.3104  2.2352  1.1354
#> -0.8822  0.1802 -0.0243 -0.4939 -2.4371  0.9546 -0.6486  0.8011 -0.3294 -0.8191
#> -1.0208 -0.3147 -0.7974 -1.8117  0.0412 -0.0038  0.8794 -1.3290  0.0292  0.0054
#>  1.4870 -1.6689  0.7169  0.8203 -0.5320 -0.1316 -0.4377  0.0965 -0.0998  0.3549
#> -2.1444 -1.3776  0.3497  1.6215 -2.8581 -0.0532 -1.4702  1.7290  1.3988 -1.4678
#>  1.6004  0.4638 -0.3930 -1.4196  0.5895  0.5343  0.9144  1.1663  1.3834 -1.0756
#>  0.0504  1.4122 -1.0623  1.9297  0.2311 -1.2400 -0.7265  0.3274 -1.7019 -1.0399
#>  0.3341  2.0121 -1.2231 -0.0735  0.1748 -0.7336  0.2212  0.5347  0.2516 -0.2005
#>  0.8667  0.8594 -0.9475  1.5689  1.0077 -0.3711  1.4350  0.6106  1.3217 -0.6127
#> -0.1769  0.3559  0.1914  0.4184 -0.5579 -0.7226 -0.1907 -0.2676  2.1859  0.3132
#> -0.2415 -0.7425  0.2806 -0.5082 -0.9638  0.5679  0.3373 -0.4162 -0.1947 -1.3652
#>  0.5929 -0.3167 -0.1077  2.1402  0.2956  1.1695  1.2290 -0.9577  0.8081 -0.2966
#> -0.6418  1.0262 -1.3926  1.8191 -0.0070 -2.0482  1.7914 -1.0242 -1.6070  0.4718
#>  0.7689  0.2136 -0.6373 -0.5707  0.2401 -0.6566 -0.2235 -1.3876 -1.0833 -2.1899
#> -0.6075 -0.0245  1.4265 -0.7436 -0.2186 -0.0044  0.5256 -1.0381 -2.5580 -0.3957
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]