Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.1659  0.4144  1.4447 -0.3316  0.3494  1.9831 -0.7556 -0.2338  1.2544 -0.8365
#> -0.5338 -1.6307 -0.7875  0.1956 -0.9885  1.1412 -0.1799 -0.2247  0.9371 -0.6517
#> -0.6065 -0.6940  1.1563  0.1665  0.0650  1.4354 -1.0332 -0.4043  0.7773 -1.4408
#>  0.1685  0.6109 -1.7123 -2.0691  2.0414 -1.0741 -0.8072  0.0609 -0.4739 -0.5341
#>  1.2438 -0.8154 -0.5555  0.3040 -1.9834 -0.3901 -0.3681 -0.6426  0.9128  1.2345
#> -0.2644  0.5661 -0.7799 -0.6747  0.1751  0.5901  0.5204 -0.3892  0.4115 -0.5073
#>  0.1876  1.3926 -0.0042 -0.5234  1.6036 -1.0908 -1.3966 -0.1143 -0.2798  1.3163
#>  0.5167 -0.5207  0.3779 -0.4185 -1.1922  0.4004  1.0358 -1.7008  0.4334 -1.1855
#> -0.2908 -0.8110 -0.0763 -0.2072 -1.3494  0.1738 -0.6823  0.2579 -0.1944  1.1846
#> -1.2853 -0.4458 -1.7565  1.0195 -1.5616 -1.1416  1.1671 -0.1334  0.5243 -1.3008
#> -0.0743  0.9006  1.1879 -0.0440  2.1978 -1.1483 -0.1214 -0.7774  0.1359 -0.8808
#>  0.5609 -1.6067 -0.6224  0.2391 -0.8045  0.1974 -0.3055 -0.7624 -1.1908 -0.5843
#> -0.0260 -0.9324  1.0561 -2.8128 -0.6816  0.1089  0.5069 -0.1796 -0.0149 -0.5330
#>  1.0506  1.3232  0.7640  0.8544 -0.5017  1.1385 -1.1322 -1.2473 -0.2367  0.1243
#>  0.4480  0.2602  0.4665 -0.1575  0.5237 -0.9060 -0.8961 -0.0176 -0.2676 -0.1472
#> -2.2278 -0.6186 -0.7166 -1.0161 -0.9642  0.7113 -1.6250 -0.4735 -0.6115  0.0961
#>  0.7046  0.7279  0.6313 -0.4751 -0.1889  0.3386 -0.2655  1.0793  0.1527  1.1066
#> -0.9539  1.4617 -0.2372  0.1487  1.2864 -0.8726  0.3156  1.5745  1.4353  0.6123
#>  0.3550  0.9483  0.1961 -0.6750 -2.0173  0.6494  1.6041 -0.2157 -1.8727  1.2440
#>  0.3765  0.6264 -0.6210  1.1423 -0.6550 -0.3553 -1.0791 -0.8673  0.8480  1.4130
#> -1.6878  1.2517 -1.4905 -0.2646  0.7627 -0.3108  0.6726  1.5779  0.7509  0.1038
#>  1.7335  0.3002  1.1784 -1.0249 -1.1134 -0.2358  1.0895 -0.2883  1.2358  0.3327
#> -0.7998 -1.2205  1.0604 -0.2907  0.0610  1.1276  0.1315  0.4821 -1.0403 -1.0117
#>  0.3051 -1.0640  0.9334  0.0109 -1.0887  2.5979  0.7112  0.7620  0.7096  1.6642
#> -0.5251 -0.3354 -1.5651 -1.1603  0.5367 -0.2005 -0.3083  2.9688  1.3058 -0.9984
#> -0.8499  0.6861 -0.2561  0.0134 -0.4121 -1.0270 -0.7833 -0.7497 -0.7011  0.5021
#>  0.5181  1.5831  0.4586 -0.7311 -0.4815  0.9987  0.5727  1.0093  1.5575  0.5648
#> -0.1277  0.3887 -0.4420 -0.8944  1.4021 -1.7179 -0.1873  2.1476 -0.0762  0.9578
#>  0.5243 -1.0930 -0.0282  0.3365  0.6953 -0.8384 -0.6257  0.5909 -0.2870  0.2956
#> -0.2497 -0.5778 -0.3173 -0.9755  0.0004 -1.7337 -1.1517  0.7089 -1.6802 -0.3940
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]