Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0876  0.1699  0.0589 -0.9377  0.4450  0.3824  0.0388  1.4332 -1.1815 -0.5244
#>  0.0727 -2.3204 -0.6365  0.3142 -0.2581  1.4826 -0.4622 -0.8126  0.8440  2.4226
#>  0.9914 -0.1474  1.1810  0.5656  0.5883  1.2420 -0.5636 -1.6132  0.8532  0.2935
#>  0.2984  1.3500  0.6726 -0.5841 -0.5485  0.7009 -0.8899  1.2069 -1.7704 -0.3532
#> -0.1770  0.2528  1.0613  1.0394  0.8141  0.9772 -1.0801 -0.7022 -0.9596 -1.0014
#> -0.6659 -0.6136 -0.1137  1.7625  0.9378  0.7898  0.3630  1.1660 -0.5271  0.9104
#>  1.1178 -0.1333  0.0774 -0.3581 -0.5678 -0.3878  1.2592  0.1027  2.6351 -0.7924
#> -1.6104 -0.4106 -0.1729 -0.3123  0.9757 -0.7194  1.8986 -1.5089 -0.2938 -0.5822
#> -0.3529 -1.3038  1.0868  0.4663 -1.2985  1.2954  0.1628 -3.0122  1.1814  0.2130
#>  0.9573 -0.5504 -0.9136 -0.7750  0.3802  2.8563 -0.6131 -0.3652  0.8957  0.5231
#> -1.7752  2.3354 -1.0561 -0.7567 -0.8126 -1.1622 -1.3074 -0.0795 -0.7660 -0.6869
#>  0.8865  1.5387 -1.1579  0.0952 -0.5425 -0.4152  0.8928  1.3852  1.2991 -1.2102
#> -0.4763  0.9749 -0.2220  0.5401  1.0847  1.3248  0.8555  0.3172  0.1355  1.0044
#>  0.9899 -1.6231  0.1099 -0.5446 -0.4479 -0.2235 -0.2648 -0.4050  0.2350  0.5132
#>  0.2632  0.5944 -1.5041 -0.0899 -0.3823  1.0176  0.0122 -0.5101 -0.0828 -0.4616
#>  0.2928  0.0703 -0.1530 -1.1282 -0.8256 -0.4476 -2.0315 -0.2162  0.9985  1.1747
#> -0.3090 -0.0507 -0.0262 -1.7865 -1.6731  1.3236  0.0877  1.0496  0.6075 -0.2935
#>  1.6164 -0.0471 -1.0428 -0.3622 -0.1984 -0.5911  1.7846  1.5652 -0.5875 -1.5281
#> -0.7170 -1.4430 -0.7497  0.9737  0.1034  1.2416 -2.9341  0.7869  0.3146 -0.0515
#>  0.7110  0.4839  1.3766 -1.0761 -1.2144  0.1097 -0.8990 -1.3510  0.1701 -0.8771
#> -0.0694 -1.1071 -1.5269 -0.6075  0.1595 -0.6603  0.1056  1.6075  1.1438 -1.9852
#> -0.0218  0.8561  1.2978 -0.7226 -0.3315  1.4406  0.4662  0.1421 -0.2366  0.3561
#>  1.9731 -1.5890 -2.4999 -0.1555 -0.5688 -0.6455 -0.3158  1.5023  1.2600  1.6394
#> -0.9288  0.9978 -0.7330 -1.6290 -0.0112  0.5845  0.8601 -1.7017 -0.0873 -0.2004
#> -1.6774  0.2398 -0.1699  0.8336  1.4521 -1.2794  1.5793 -1.2188  2.4929  0.6960
#>  0.2787 -0.1871 -0.4716  0.7006 -1.0481 -0.5697  0.4248 -1.6359 -1.2902 -0.5779
#>  0.6972  0.9333 -1.0931  0.2826 -0.9037  0.0079 -0.1010 -0.6419  0.4156 -1.2545
#> -1.6859  0.2119  1.5325  0.8681  0.3147 -0.6322  0.2696 -1.5735 -0.6266 -0.5729
#>  0.4506  0.7168  0.6531  0.9382  1.4907 -0.4469  0.2352  1.2725  0.7150  1.5042
#>  0.1959 -1.3456 -0.2721 -0.5882  0.3245 -1.1931 -0.4119 -1.4272 -0.2508 -0.4653
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]