Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7511 -0.4697 -1.0513 -0.7469  0.2269 -0.5025  0.9576 -1.8613  0.1198 -0.0283
#>  0.0363 -0.2139 -0.9913  0.6578  1.3670  0.8606  1.1737  0.6280 -0.0768  0.5243
#> -1.9015  1.6749  0.1189  2.6857  0.8606  0.1230 -0.4793 -0.4306  0.2234 -0.0662
#> -0.6786  0.5980  0.5594  1.6936  0.4019 -0.0424 -2.4132  0.5101 -1.3258  1.1129
#>  0.1664 -0.6835 -0.1832 -0.1307 -0.5550 -1.5047 -3.1073 -1.0512 -0.0907  0.4285
#> -1.1823 -0.0533 -0.2315  2.5515  1.0463 -2.1329  0.1013 -0.2936 -1.0183 -1.7226
#>  1.3089 -0.4738 -0.1580  0.3407 -0.4859  0.0528 -0.8506 -0.8694  0.5374 -0.6181
#> -1.0651  3.2410  0.9191 -0.3730 -0.1950 -1.5881 -0.1557  1.2729 -0.3735 -1.2002
#> -0.1998  1.0637  0.0792  0.6262 -0.1033 -0.7553  0.1575 -1.7243 -0.6259  0.0522
#> -1.2720 -0.1946  1.6579  0.0914 -0.1299  0.9297 -1.4355 -0.6122 -0.4955 -0.7492
#>  0.7268  0.1149  0.1183  0.5750  0.0091 -0.7807  2.1393 -0.9728  0.6488  1.2127
#>  1.3676  1.1300 -1.2027  0.7977 -0.1957 -0.0232  0.2339 -0.7996  1.1708  0.8837
#> -0.5291  1.3172 -1.8848  2.0291  2.1901 -1.7995 -0.1688  0.1977 -0.3928  1.1947
#> -0.4806 -1.0090 -0.3816  1.9430  0.1089 -0.2320  0.5123  0.4575 -0.4293  0.1190
#> -1.0844  1.8066 -0.6384 -0.4716 -1.4367  0.4407 -0.0666 -1.0199 -1.0156 -1.5372
#> -0.1032 -1.0357  0.2262 -2.1042  0.6233  0.9483  2.4804 -0.0450 -0.4916  2.5611
#>  0.1826 -0.5427  1.6808  1.5891  0.3323  0.4410 -0.0797  0.7718  0.3933 -0.9583
#> -1.2093 -1.1246 -0.2918  0.5512 -0.8323  0.0791 -0.4382 -1.9496  0.7089 -0.3475
#> -0.3776  0.8757  0.5123 -0.4063 -0.1798 -0.1977 -1.1331  0.1162  1.6868 -0.1909
#> -0.8327 -0.3705 -1.0904  0.7154  1.3934  0.3759 -0.2359  1.3517 -0.7744  0.8712
#>  1.9660 -0.2230 -0.6161  0.1364  1.4593 -0.3803  0.8931 -0.2778  2.7220  0.0502
#>  1.6379  1.5730  0.6439  0.0827 -0.1297  0.7719  2.1813  0.2019 -3.0898 -0.6890
#>  0.1195  1.5490  0.2632 -0.9345  0.5687  0.7528 -0.8097  0.9113  1.0724  1.1747
#>  1.7968  1.6714  0.0409  1.0705  0.5550 -0.1689  1.2603  0.2112  1.1816  1.6259
#>  0.1062  1.1772  1.3654 -0.5791  0.4175  1.0938  1.0795  0.7933 -0.9597  1.5163
#>  0.5991  0.8045 -0.4035  0.7920 -0.6413 -1.4008  0.1355 -0.5656 -1.7215  1.5568
#>  0.3650  0.6464  1.5826 -0.0905  1.9336 -1.0441 -1.3414  0.0881 -1.7485 -0.8268
#>  0.9924  0.1421  1.8488  1.2334  0.1375 -3.1268  0.8994 -1.1484  0.0651  0.8277
#> -0.8161  0.6327  1.0556 -1.9058  0.6031  0.1157  0.2962  0.1215 -0.8524  1.9236
#>  0.4377 -0.8683  0.5869  1.2926  1.4312 -0.7477  0.1164 -0.3982 -0.1906 -0.1536
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]