Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3939  0.7657 -1.5416  0.9600  0.3496 -1.5996 -0.7562  0.8211 -0.7034  0.5771
#> -0.3694  0.2616  0.3609  0.0720 -0.1049  0.4171  1.1796  0.3773  0.9136  0.9994
#>  1.9253 -1.7194  0.4251  0.1011  1.0233 -1.1082  0.8897 -0.8595 -0.4621 -1.7206
#> -0.0694 -1.3972  0.2214  2.3839  0.0760 -0.0929  1.2535 -1.6210  0.9722  2.1577
#> -0.0402 -0.0972 -1.5536  1.6651 -0.9080  0.5049  0.3041 -1.2160 -1.4066 -0.7001
#>  0.5609  1.5954 -0.2237 -0.8836  1.4409 -0.1734  0.2617 -0.3199  0.0545  0.3055
#> -0.1824 -0.6233  1.1443  1.5782  2.0903  0.4672  1.3589  1.3128  0.4254 -0.8523
#> -0.2634 -0.7898 -0.4081 -0.2387 -0.7657  1.2499 -1.0874  1.3345  1.7328 -1.2144
#> -0.5384  0.9371 -0.5621 -0.7637 -0.0559 -0.3074  1.0430  0.0113 -0.6811 -0.8034
#>  0.9003 -0.0730  0.3841 -0.1469 -0.4333  2.1446  0.9397  0.5014  0.8541 -0.3799
#> -0.2212 -0.5949 -0.5350  2.0495  0.7724 -1.5151 -0.8107 -0.8814 -2.3826 -0.4649
#> -0.0028  0.3246 -1.5537 -0.2039  0.1695  0.8601  1.6639  0.5318 -0.6062  1.8430
#> -0.0306 -1.5449 -0.1718  3.2739 -0.4070  0.2029 -0.6273 -0.3671 -0.8935 -1.0692
#>  0.7870 -0.1615 -0.8877 -0.7042  0.6410 -0.2505  0.1260  0.0742 -0.6970  2.1470
#> -0.2778 -1.6857  0.5710 -0.1012 -0.9999  1.0381 -0.1586 -0.6507 -2.0007  0.1496
#> -0.7050 -0.1003  0.3275 -1.5812  1.2514  1.3550 -0.3331  0.6804 -0.3674 -0.2568
#> -0.7777 -0.4153  0.2384 -0.6192 -0.6534  0.5588 -0.3016  1.7597 -1.1592  1.2332
#>  0.2928  0.0993 -1.8396 -0.0009  0.6661  1.4400 -1.6060 -0.7514  0.0483  1.3304
#>  1.2331 -0.4019  0.6156  0.8485  0.9459  0.1906 -1.2123  0.2597 -0.2327  1.4990
#> -1.0480 -0.1090  1.2299 -0.5502 -0.0807 -0.4454  1.2733  1.9824  0.3335 -1.5432
#>  0.7545 -0.3543  0.1800 -0.2489  0.4559 -0.3366 -0.5035  1.0026 -1.2736  1.4473
#>  1.4481  1.0495 -0.0666  0.9227  0.4479  0.5312  0.3094  1.3901  0.6621 -0.5769
#>  0.0231 -0.0134  0.7371  0.7794  0.4378  0.7583  0.6993 -1.4839  1.5222 -0.6125
#>  1.8286 -0.7561  0.6095  0.2299 -0.8254 -0.6474 -0.0086 -0.0130  0.1974  0.2959
#> -1.3407  0.5529  1.2512 -0.0081  0.5686 -0.7970 -0.5587  0.5925 -0.9863  0.4561
#> -1.7880  0.3982  2.6824 -0.1576  0.5753  0.2307  2.0876  1.0516  0.8383  0.3326
#>  0.7780  0.6413  1.0170  1.7898  1.0365  1.2546  0.8588 -0.6856 -0.3073 -0.8463
#>  0.4255 -1.6543 -0.6546 -0.2575  0.7139 -0.6784 -1.3726 -1.5525  0.3502 -0.0543
#>  0.1498  0.1717  0.9801 -0.4679 -1.9784 -0.1445  0.3324 -1.7420 -0.9328  1.5104
#> -0.0123  0.7656 -0.6085  0.8403  1.5618 -0.1675 -0.1570 -0.2262 -1.2160  0.4702
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]