Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2183  0.0353  0.2039  1.4135 -0.6470  0.7975  0.1070 -0.7557 -0.7332 -1.8277
#> -0.6840 -1.4316 -0.0013 -1.3220 -0.8530  0.1417  0.1236  0.1501  0.3129  0.7971
#> -1.4299  0.5672  0.1600  0.7793  2.6442 -0.6062 -0.3490  0.2210 -0.3491 -1.1764
#> -0.7049  1.8628  0.5710 -1.0807  0.7315 -1.7549  1.2544  0.4632 -0.1522 -0.3687
#> -0.4934 -0.8215  1.7408 -0.5961 -1.4031  0.4326 -0.9790  0.6300  0.0108  1.3000
#> -0.0506  0.3489  0.9059 -0.3213  0.9168 -1.1428 -1.2903  0.9181  0.3933 -0.6307
#> -1.4525 -1.8678  0.5947  1.4862  0.7874  1.1878 -0.4825 -1.5568 -1.1924 -0.0739
#>  0.6571 -0.3188  0.4249  0.2811 -0.7243 -0.3570 -1.0975  2.2954 -1.1674 -0.4734
#> -0.7352  1.1719  1.3686 -0.9329  1.0858  0.0685 -0.7841 -1.2306 -0.7693 -0.5568
#>  0.3232 -0.7063 -0.3449  0.3241 -0.0749 -0.5823  0.1487  0.3798 -0.8244 -1.3134
#> -0.5229  0.7023  1.3544 -0.6211  1.5490  0.3637 -1.7876  0.1521  0.5917 -1.4688
#> -0.3377  0.3136  0.1363 -1.1060 -0.0369  0.6392 -0.1128 -0.8782  1.2158  1.4149
#> -1.3592 -0.3794  0.1498 -0.2392 -0.3484 -0.8411 -0.4630 -0.6838 -0.4451  0.6465
#>  0.5506 -1.1206 -0.8451  0.6786 -1.0830  1.2713 -1.3214 -0.4673 -0.7841 -0.3146
#> -0.2347  1.5012  0.7278  0.7400  1.1456  0.4612 -0.3413 -3.0133 -0.1418 -0.6143
#>  0.9169  0.5260  2.5714 -0.2689  1.1763 -0.6116 -1.1042  1.0609  0.5146 -1.4839
#>  0.5327  0.1842  0.6527  1.0254 -1.2175  0.1553  0.0792  1.7948 -0.5617 -0.5541
#> -0.2795  0.9878 -2.4989  1.0145 -0.2746  0.1810 -0.6808  0.5654  2.0582 -0.2236
#> -1.2037 -0.9830  1.6182 -1.3483  0.8722  0.1467 -0.9358  1.4072  1.8799  2.1849
#> -0.6573 -0.6827  0.1322  1.4120 -1.6432 -1.6041 -0.5201 -0.8167  0.7782  0.2629
#> -1.1879  0.9988 -0.3848  0.3660 -0.3213 -0.5321 -0.9120  0.2360  0.0851 -1.3739
#> -1.0952  2.4712  1.1101 -0.5860 -0.3999 -1.3834 -0.3542 -0.7925 -0.5990 -0.1296
#> -2.0676  0.7555 -1.0151 -0.3620  0.4853 -0.3984 -1.5047  0.4917  1.8577 -0.9199
#> -0.3665 -0.2068  1.6119  0.2051  0.9682  1.0930  0.1140  0.4231 -0.8375 -1.9137
#> -0.0334 -1.2234  1.3294  0.2651 -0.2789 -0.1450  1.6917 -1.5024  0.7456 -1.7463
#> -1.2106  1.1758 -0.0523  0.0445  1.3124  0.8627  1.3001  1.4847 -1.2367 -0.3077
#> -0.9935 -0.5129 -0.5519 -0.4179 -1.0091 -0.2053 -1.2760  0.2108 -0.5604 -0.7181
#> -0.7653  0.4591 -1.6038  0.5971  2.6961 -0.0523  0.4400 -2.0454  1.1612  1.0566
#>  1.4800  0.9905  0.6708 -0.0894  1.2996  0.3681  1.5412  0.8558 -0.1429  0.3825
#> -0.5109 -0.0315 -0.9903 -0.3222 -0.5600  2.1141  0.0887  1.4859  1.0456 -0.4836
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]