Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6-4.0386e-01  1.2875e+00 -2.9316e+00 -5.7278e-02  1.1655e+00  7.5149e-01
#> -1.6737e+00  2.7906e-01  9.1715e-01 -1.2095e+00 -6.9219e-01  1.8856e+00
#>  1.5531e+00 -4.2347e-01 -1.3695e+00 -1.8054e-01 -7.3532e-01 -1.7995e+00
#>  1.3762e+00 -1.4981e+00 -1.3762e-01 -6.0487e-01  1.4000e+00 -7.3993e-01
#> -3.3698e-01  1.4382e+00  1.9421e+00  7.1845e-02 -3.8039e-01  1.2701e-01
#> -3.2930e-01 -1.2715e+00 -6.9248e-01  1.0063e+00  5.1368e-01 -9.6437e-01
#>  9.3739e-01  7.6977e-01 -1.2850e-01  1.0606e-01  4.1938e-01  1.8009e+00
#> -6.6465e-01  6.0221e-01  2.4279e-01  3.9739e-01  4.5804e-02 -3.5537e-01
#>  1.9032e+00  6.1917e-01 -4.9699e-02  7.0002e-01  9.6035e-02  7.9673e-01
#> -9.1877e-01  9.8924e-02  2.2011e-01  6.4756e-01  3.0455e-01  6.9976e-01
#>  4.8072e-01 -7.5266e-01  1.1020e-01  8.9751e-01  8.9021e-01 -1.8332e+00
#>  2.2467e-01 -4.8247e-01  1.4057e-01 -1.9503e-01  1.4959e+00  6.3474e-01
#> -3.4600e-01  7.6256e-01  2.3301e-01  1.4719e+00  1.9465e-01  3.0964e-01
#>  7.2807e-01  4.6680e-02 -1.7213e-01  2.1647e-01  5.7502e-01  1.8961e-01
#>  1.0109e-01  3.4038e-01  8.6331e-01 -9.0424e-01  6.5556e-01 -2.3356e-01
#>  1.2108e+00 -3.2632e-01 -5.4355e-01 -9.8771e-01 -8.1262e-01 -1.1054e+00
#>  1.3854e+00  1.1839e+00 -2.7520e-01  4.9627e-01  9.3733e-01  7.7263e-01
#> -1.1679e+00  2.1562e-01  3.3792e-01  2.2474e+00  9.7164e-02  3.9847e-01
#> -1.3998e+00 -6.8769e-01 -3.8785e-01  1.0495e-01  1.3543e-01  4.5720e-02
#>  4.2590e-02  4.8894e-01 -8.4155e-01 -1.0279e+00  6.6483e-01 -6.3531e-01
#> -5.3425e-01  1.2941e+00 -8.8226e-01 -9.6982e-01  9.9813e-01  3.5164e-02
#>  1.7336e+00 -1.2762e+00  5.9493e-01  7.9375e-01  6.9699e-02  9.4321e-01
#> -2.3093e+00 -1.6817e-01  8.7462e-01  4.7798e-01  6.3480e-01 -7.1193e-01
#>  8.2025e-01  1.0886e+00  1.0035e+00  1.2658e+00  1.8172e-01  9.4923e-01
#>  7.3355e-01  8.7026e-01 -8.9192e-01 -1.1519e-01  5.1628e-01  1.0633e+00
#>  1.0215e+00 -4.4654e-01 -3.8488e-01 -1.3723e+00  4.1321e-01 -1.5571e+00
#> -6.9102e-01 -3.1661e-01 -1.5103e+00  4.8736e-01 -1.5894e+00  5.4790e-01
#> -2.8524e-01  1.6872e-01 -2.3067e-01  4.2543e-01 -8.0730e-01  1.1100e+00
#> -7.5056e-01 -1.0047e+00  4.0556e-01  6.1480e-02 -1.3441e+00 -3.7885e-01
#>  2.5423e-01  1.8811e+00  2.1000e+00  2.2022e-01 -1.0311e+00 -3.3338e-02
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]