Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.2222 -1.4102 -1.7168 -0.0292  1.4162  0.3507  0.8818  0.6514  1.5648  0.9969
#>  0.9496 -0.3244  0.8534 -1.5500  1.1908 -2.9446  0.1158 -1.0968  0.9536 -0.5914
#> -0.6600 -1.6354  0.1112  0.9375  1.5841  0.2745  0.9867  0.4558  0.1348  1.5985
#>  0.3407 -0.6519 -0.6560  0.5317  1.6016  0.0121 -0.2088 -0.2573  1.0978  0.2427
#>  0.6642 -0.4992  0.5843  0.6922 -0.5744  0.7894  0.5051 -0.8708 -2.5022 -0.5604
#>  0.1827  0.3664  0.1561 -2.3446  0.2308 -0.3130  0.0259 -1.4074  0.5342  1.3558
#> -0.3253  0.1350  1.1677 -0.8371 -0.3384 -0.8390  0.3364 -0.0567 -0.3587  0.6714
#> -1.0372  1.8394  0.3172 -0.3136 -1.0791 -0.3395 -0.7442 -0.9273 -0.7421 -1.2850
#>  1.5583  1.9691 -1.2470 -0.1377  0.5089 -1.2794 -0.6356  0.0129  1.6891  0.5427
#> -0.6828 -0.9042  0.5839  0.2140  1.3245  0.2996 -1.7111 -1.8846  0.2621 -0.6056
#> -0.1890 -0.9832  0.6270  0.0543  0.8150 -0.2274 -0.9524  0.1918 -0.5885  1.3064
#>  0.0503 -0.3681 -1.0247  0.0870 -0.6117 -0.1162 -0.2933 -0.3449 -0.3200  0.1931
#> -0.2766  0.8032  0.3427  0.2746 -0.8287 -0.6356  0.1698 -0.0948  1.3736  0.5711
#> -3.8733 -0.7151  1.4138  0.2915 -0.5288 -0.3270  0.3385 -0.2229 -1.3399  0.0380
#> -0.2285 -1.9378  1.0737 -0.4161  0.8977 -0.1257 -0.4112 -0.0650  0.9384  1.2514
#>  1.0739  0.4495 -0.4715  0.6609 -0.6660  0.7640  0.4387 -0.5568  0.5112  0.1279
#> -0.2577  3.0304  1.1138  1.2583 -0.6364  1.6717 -2.6012 -1.5597 -0.7029  0.0309
#>  1.0099  1.4563  0.7345 -1.5982  0.3907 -0.2505 -1.2761 -2.0695  0.3849 -0.4693
#>  0.9444 -0.1824  0.4645  0.4656  1.1634 -0.1625  0.0837 -0.5410  1.4143 -1.4723
#>  0.4383 -0.6322 -0.0236  0.3260  0.6612 -0.0076  0.3753  0.0224 -0.1322 -0.4244
#> -0.3775 -1.6422  0.8204 -0.2970 -0.4062 -1.6466 -0.5871  0.1419 -2.5634  0.1848
#>  0.5844  0.3918  1.2036 -0.2853 -1.5474 -0.9542  0.8304  0.9953 -0.7074 -0.3235
#>  0.8932 -2.5731  1.1669 -0.0089 -1.1774 -0.7546 -0.9877 -1.1914 -1.0423 -0.8073
#>  1.6081  1.3485 -0.7274 -0.1906 -0.1111 -2.8489  0.3486 -1.6774 -0.8099 -0.5770
#>  0.9580 -0.5030  0.0225  1.4359  0.1239 -0.1342  0.4224 -0.3903  0.9667  0.1898
#>  1.5249  0.7326  1.1205  0.9062  1.3432  0.0159 -0.7370 -1.3815 -0.3084  0.6379
#>  2.0622 -0.8052 -1.3308 -1.4945  0.2458  1.5157 -1.2151 -0.9187  0.0926  1.0570
#> -0.0488 -0.1941 -1.5486 -0.5371 -0.0424  0.5136 -0.4858 -0.2182  0.1427  0.8169
#> -0.0027 -0.7638  2.2775 -0.5117 -0.4407  2.0109 -0.5888  0.0744  0.0849 -0.6328
#> -0.0184 -0.6807 -0.5897  0.7872  0.3614 -0.6100 -0.0275  0.3304  1.0415  1.6612
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]