Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.1381 -1.2290  1.1293 -2.7699  1.6639 -0.7058 -1.2673  0.0749 -1.1561 -0.5436
#> -0.3719  0.4178  0.9569  0.3584  0.1595 -1.0260  0.7729  1.3829  0.3780  1.6666
#>  0.3580  1.3712 -0.1422 -0.0026 -1.0434 -1.2104  0.3070 -0.0591  0.0921 -1.2770
#> -0.0471 -0.6472 -0.1803 -2.0677 -1.1677 -0.9254 -0.4832 -0.2986 -0.7453  1.0410
#> -0.6582  0.0755 -0.5451  0.7651  0.3192 -0.9785  0.7466  1.0900  0.5899  1.6030
#>  0.5428  0.3072 -2.1105 -1.1572  1.2198  0.7431 -1.5823 -0.3710  1.5938 -1.2362
#> -0.7962  0.0106 -0.4167 -0.1391 -0.6498  0.0466 -0.2725  0.9852 -0.9608  0.0831
#> -1.2055 -0.7108 -0.7756 -0.9793  1.4189  0.6618 -0.5722  1.0912  0.0286 -1.2192
#> -0.3274 -0.2291  0.3145  0.4893  0.6081 -2.1216 -1.1221 -0.6028  1.3723  0.2005
#> -0.4992  1.0222 -0.0121  0.2462  0.5098  0.4559  0.2636 -0.9484  1.7551 -0.8270
#>  0.0132  1.4367 -0.5920 -0.3199  0.1974 -1.3533  0.1508 -1.3103  0.8215  0.2891
#> -0.2318  1.6089  1.4433 -0.1635 -0.7760  0.5613 -0.6812  0.2005 -0.7042 -1.6767
#>  0.7828  0.7071  0.6646 -1.9148  0.5113  1.5861  0.0824  0.3883 -1.4101 -0.2946
#>  0.2718  0.8540 -0.7600  0.1277 -0.8673 -1.3523 -0.5696  0.5756 -0.5628 -1.4933
#> -0.5938  0.4742  1.8978 -0.1429 -0.2021  0.6882 -1.2546  1.5710 -1.2878  0.5532
#> -0.0936  0.0996 -0.9446  1.5157  0.3656  0.9348  0.4547 -0.2137  1.2408 -0.0658
#> -0.5350 -0.4302 -0.6467  0.3455  0.4371 -1.8666  0.2424 -2.1358 -0.0242 -0.0715
#>  1.5172 -1.6897  0.5334 -1.4721 -0.1616 -0.6914 -0.2669 -0.5604  0.9193 -0.1615
#>  0.5686  0.0954 -0.8729  0.7452 -0.1386  1.4869 -0.4113 -0.7517  1.0794  1.0152
#> -2.4750 -0.7100  0.6682 -0.3075  0.6393  0.3027 -0.0350 -0.0965 -0.5335 -0.6441
#>  0.2138  0.4894  0.7837 -1.0638  0.4076 -0.0946 -0.8652 -1.3212  0.2886  0.7583
#>  0.7036 -0.6122  0.3887 -1.4568  1.5264 -1.2535 -0.7467 -1.2258  0.8545  0.0242
#> -1.0107  0.8819  1.0750  0.6099  0.2837 -0.3120 -0.5519  1.1737  0.6364  0.4246
#> -0.1684  1.4820 -0.9105  0.1822  0.5067 -1.3839 -0.9190  0.8839  0.8112 -1.7346
#> -0.7636 -0.3709 -0.0099  0.0119 -0.3758 -0.8071 -0.1379  1.4360  2.3933 -0.3444
#> -1.2035  0.4379 -0.9197 -0.5106 -1.3504  0.8700 -0.2943  0.8495  0.7083 -0.8534
#> -1.9159 -2.1068 -0.2540  0.0392  0.4216  0.6727 -0.2597  1.3962 -1.8350 -1.3151
#>  0.0787 -0.7676 -0.0955 -0.2500  0.2348 -0.6096  1.3565  0.4415  0.7521 -0.1863
#>  0.2400 -0.5983 -0.6434 -1.2000 -0.7097  0.1164  0.2763 -0.8088  0.6536  1.4285
#> -0.6557  1.4274  0.4955  1.8207  0.8473 -0.8755  0.7065 -0.5350  1.2515  0.3811
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]