Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.4930  0.2138  0.2225  0.4565 -0.1021  0.6828 -0.1785 -0.2993 -0.6238  2.0870
#>  0.4635 -0.2274 -0.1948 -1.2525 -0.0921  1.3750  1.2153  1.0352  0.5970 -0.2595
#>  0.5002 -0.3459  1.1887  1.0213  0.5098 -0.9033 -2.0543  0.2735  0.9606  0.0297
#>  0.1051  0.6626 -0.0551 -2.0552 -0.7841  0.6711  2.4235  0.6383  0.8737  0.5135
#> -0.5578  0.8374 -0.5886  0.2036 -0.7108  0.0285  0.0235  0.7295  0.1371  0.1533
#> -1.6245  1.7335  2.1739  0.4357 -0.5347  0.4634  0.8214 -0.4969 -0.6250  1.1612
#> -1.2334 -1.2400  0.9215 -1.3054 -0.8851  0.5802  1.4047 -0.9543  0.0789  1.9067
#>  1.2138  0.3536 -1.3878 -1.1497  0.8421  0.7248  1.6219  1.5900 -1.0005 -0.4509
#>  0.2921  0.3108  0.8994 -1.1808  1.6513 -0.1888  0.1309 -1.7205  0.2441 -1.4360
#>  1.5877  1.5214  1.6745 -0.1799 -0.7864  1.8243  0.1267 -1.1173 -1.1350  1.4539
#> -1.1429 -0.3961 -0.1204 -0.2708 -1.9101  1.9125 -0.2712  0.0134  1.3065 -0.4198
#>  0.5245  0.8981 -0.1544  1.2568 -0.0062  1.6351 -0.6597  0.4345 -0.4882 -1.0910
#>  0.2669 -0.1537  1.5884  0.2781 -1.0839  0.0639  0.2845 -0.7285 -0.7607  0.3205
#> -1.0778  0.8668  0.9946  3.1201 -0.5276 -0.1833  0.7741 -0.5693 -0.2554  0.0250
#> -0.5960  0.7990  1.4003  0.1240  0.7122 -0.4426 -0.4646 -0.2287  1.3740 -1.5425
#>  0.9660 -1.0735 -0.2766 -0.6415 -2.4024  0.0038 -0.2178  0.1639 -0.0411 -0.0743
#> -0.8338 -0.3613 -0.2581  0.8990  1.3146 -0.6897  1.6677 -0.1908  0.4322 -0.3832
#>  1.1319  0.3083 -0.0475  2.1025 -0.7594 -1.2080  0.5594 -0.0681 -0.3214  0.4718
#> -0.3692  0.6935 -0.6003 -0.3886 -0.8731  0.2040 -0.5262 -1.1489  0.6867  0.0406
#> -0.9571  1.9147  1.5596  0.1156  2.6886 -0.9808  2.1569  0.0772  0.2739  0.2020
#>  0.4519  0.5417 -0.4430  1.5200  0.7694  0.4548 -0.3111  1.6500  0.0383 -1.6427
#> -0.7867  0.2790 -2.2983  1.9742 -2.3392  0.9350 -0.2964 -1.4158  2.6144 -0.6728
#>  0.5163  1.4118  1.1941 -1.4200 -0.2077  1.7378  0.2680 -0.5777 -0.4374 -1.1577
#>  0.7245 -0.2015  1.3807 -1.3780  0.1297 -0.5465 -0.9372 -0.7531 -0.6786 -0.6638
#> -1.5723  0.9269 -0.2058 -0.3453  0.0694 -0.7339  0.0869 -0.0583 -0.4236  0.2097
#> -1.5741  1.6721 -0.0363  0.3953  1.0753  0.9475  0.7258 -1.1385  0.3466 -0.4747
#>  0.1664  0.2420 -0.2242 -0.1574 -0.5034 -0.3733 -0.1029  0.6592  1.8718 -0.0981
#>  0.8001  1.5892 -1.5084  1.0709 -0.9080 -1.4024  1.0563  0.4001 -0.3092  0.0661
#> -2.2748  0.9798 -0.9560  1.4384 -0.8477 -0.7082 -0.7070  0.0482  0.5699 -0.2947
#> -0.3232  0.2758  1.2047 -1.0211 -0.0203 -0.1484 -0.2450 -0.6891 -0.4676 -0.7634
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]