Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0633 -1.1577 -0.5366 -1.9304 -0.1177 -1.2472  0.3953  1.0104  0.9826 -1.4805
#> -0.6963  0.5419 -1.2188 -0.4224 -1.6771 -1.8306  1.3173 -0.9106 -0.5251 -1.0071
#> -0.5653  0.3418 -0.2722  0.2526 -0.0753 -0.8455 -0.1936  0.4541  1.4213  0.1992
#> -0.1634  1.5044 -1.3573 -1.1829  1.0250  0.2743 -1.2816  1.1114 -0.4026  0.2601
#> -0.9853  0.0649  1.2149 -1.1520  1.6224 -0.9715  0.7445  0.4354 -1.1299  1.0530
#>  0.9963 -1.1401  1.2005 -0.7771  0.2447 -0.0930 -0.4010  1.0556 -1.1973  1.2696
#>  0.5513 -0.8962 -0.4316 -1.0204  1.2406 -2.2073 -1.1704  0.3832  1.1009 -1.0325
#>  0.5565  0.7770  0.7083  0.9033  0.8189 -1.3134 -0.1109  1.2800  1.2658  1.2712
#>  0.0012  0.7148  0.1135 -1.7837  0.1801 -0.2384  0.3291 -0.2597  0.4370  0.4842
#>  2.0469 -1.5335  0.2582 -0.7951  1.5921  1.1610 -1.7153 -0.8411 -0.2905  0.1285
#>  0.9949 -0.4501 -0.2708 -1.1565  1.7816 -1.1353 -1.9182  0.7307 -0.8081  1.6195
#> -0.7135  0.7379 -0.8866 -0.9772  0.4669 -0.7037 -0.6062  0.1835  1.0935 -1.0689
#>  2.2370 -0.6870  0.6085  0.5396  0.1696  0.4722  0.2782  0.0084 -1.2223  0.9150
#> -1.0529  2.9380 -0.4136  0.6641  0.8701  0.1265 -0.8488  0.3718  0.6157  1.1691
#>  1.2814 -1.2101 -0.0087  0.9321 -0.2831  0.0695  0.3819  1.3642 -0.1326 -1.7057
#>  0.4829 -0.8539  1.1090 -1.0593 -0.8774  0.8953  0.9549  0.7170 -0.1448  0.7802
#>  1.3836  0.6708 -1.4576  1.2734 -0.3385 -0.3966  1.1792 -1.3438 -0.1962  0.1401
#> -0.5093  0.3958 -0.5227  0.4318 -0.1837 -1.1427 -1.5658  0.4484 -2.5176  0.0862
#> -1.5753 -1.1920  0.7548 -1.1172 -0.1210  0.2562  1.3883 -0.4619  1.8982  0.1048
#> -1.0962 -0.2355  0.9236  1.1687  1.7408 -0.0352  0.8527 -1.2795  0.7995  0.9970
#>  0.4902 -0.5083 -0.6552  0.6713 -0.5117 -0.8867 -0.0213 -0.3034 -0.9960 -1.1843
#> -0.0740 -0.1036 -1.7116 -2.1604  0.1657  1.2541  0.9964  0.2623 -0.9815 -2.6998
#> -1.0373 -0.9588 -0.2542 -1.4371 -1.8421  0.3378 -0.8116  1.3158  1.4390 -0.5636
#> -1.2436 -0.5906 -0.1656  0.4427  0.8182 -0.2236  2.4692  0.6656  0.5954 -0.5175
#> -0.9153 -0.2360 -0.6704 -0.2372  0.2775 -1.8783  0.5167  0.1624  0.8750 -0.0551
#> -0.1990  1.4167  0.8699 -0.2727 -0.0963  0.3856  0.3604 -0.0671  0.1011  0.6320
#>  0.4947  2.2716  1.1165  1.8692  1.2780  0.4483 -0.0908  0.7853  0.4318  0.1154
#> -0.7648  0.4339  0.0431 -1.6063  0.3453  1.0232 -0.0537 -0.2414  0.4748  0.4104
#>  0.6444  0.6588 -1.0353 -0.7211 -2.0638  0.0605 -0.9867 -1.4186 -0.2658 -0.1106
#>  0.4736  0.5677  0.2421 -1.9285 -0.1262  1.7476 -1.6736 -1.1928 -0.4171 -0.6380
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]