Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.8534  0.4194  0.3374  2.1091 -0.7120  0.1307  0.2345 -0.0078 -1.1141 -1.9591
#> -0.7672  2.4531 -0.6310 -0.6304  0.3784  0.5837  0.3556  1.5167  0.1043  0.7531
#> -0.1479 -0.2722 -0.9429 -0.8903 -1.3260 -0.3901  0.4298  0.6472  0.4495  1.6101
#>  0.6550 -0.7978  1.2351  0.6191  0.1653  0.4317  0.8386  1.3713  0.9187 -0.1544
#>  0.1669 -0.2915 -0.0465 -0.3566  0.9697 -0.6696 -0.7683  0.1449  1.0245 -0.5360
#> -0.1347 -0.4110 -0.2928 -0.2918  0.4695  0.6263  0.9893 -0.8341 -0.9384  0.2511
#> -0.8545  0.6084  0.2293  0.0931  0.0865 -0.1957  0.0814 -0.6437  0.3686  0.8876
#> -0.4879 -2.5748 -0.7139  0.6432 -0.8816  1.7305 -0.5811 -0.1113  0.0444  1.4758
#> -0.8366  0.9357  2.4770  1.2292  0.2659  0.4244 -0.8826  1.5992  0.5072  0.2769
#>  1.3621  0.3896 -0.9527  0.3018 -1.0213  1.2605 -1.1754 -0.6523  0.2308 -1.2073
#>  0.6170  0.0443 -0.7981  0.1326  0.5257  0.0086  2.5355  0.4564 -0.4222  0.3466
#>  0.6307  0.4151 -1.2890  1.6181  2.4471  0.7083  0.4785 -1.0500  1.8038  1.0339
#> -0.1199  0.3387 -1.9841  1.0633  1.6186  0.0626 -0.3479 -2.1829 -1.3150  1.3037
#>  0.2877 -0.0094  0.3158 -0.5518  0.3648 -1.5665  0.1422 -0.0085  0.6841 -0.9048
#>  0.6921  0.0658  0.5288 -0.3389 -0.0617  0.9120  0.8992  0.8758 -0.2110 -0.1473
#>  0.7336  1.5240  0.9220 -1.0715 -2.1214 -0.3355 -1.7789 -2.0005 -0.1185  0.0363
#>  0.1894 -0.0527  0.4098  0.5919  0.6538  0.9395  0.8548  0.6297  1.4419  0.3567
#> -0.6557 -0.0474 -0.1448 -1.0719 -0.2382 -0.1546  0.7339 -0.9748 -0.9686  0.7447
#>  1.5569 -1.6901  0.1430 -1.4858  0.8829  0.1338  0.9361  1.1814 -0.6249  0.4676
#>  0.6643  0.6297 -1.8444 -0.4313  1.5233 -0.1156 -1.0692  0.4722 -1.9253  0.2592
#>  1.3130 -0.0916 -0.4783  0.2526 -0.6617  0.6050  1.9431 -0.6521 -0.6476  1.2809
#> -0.0769  0.6110 -1.4955 -0.3340 -1.5830 -0.4411  0.2150  0.1890  0.5008 -0.8346
#>  2.0574 -0.7112 -0.3850  0.6334 -1.4828 -0.1826 -1.1185 -0.4688 -0.7008 -1.7260
#> -0.8985  1.4716 -1.0669  1.2208 -0.5306 -0.0030  0.6942  0.3461 -0.3952 -0.6390
#> -1.1942  0.4167 -0.5370 -1.1700 -0.3424  1.6535  0.5193 -1.3973  1.3420  2.6875
#> -1.0110 -0.2911  2.0864  1.0394  0.4037 -0.9677  1.5498  0.9613  0.5507  0.7668
#> -2.2325 -1.2186 -0.0464 -1.4015 -0.0200 -0.5183 -0.9521 -0.5782 -0.9548 -0.8466
#> -0.4633 -1.0733 -0.4401  0.2588 -1.3512  0.7867  0.5249 -0.2866 -0.8486  0.0472
#> -0.4250  1.6891 -0.4998  0.6756 -0.0986  0.0663 -0.4145 -0.0269 -0.3638 -0.8209
#> -0.1864 -1.2615  0.2230 -0.5122  0.3493  2.3992 -0.3502  0.7281  0.1231  1.3053
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]