Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3264 -0.9429  1.1344 -1.7205 -0.2821  0.7036 -1.7314 -0.2937 -0.8152 -0.4468
#> -0.5169  2.5577 -1.3033  0.2650 -1.6583 -0.1576  0.6459  0.8058  0.9563  0.6184
#> -0.1521  1.6432 -1.2061 -0.5002  0.2248 -0.9310 -0.9581 -0.1485 -1.2495  0.2876
#>  0.9385  1.1171  0.6741 -1.2955  0.8457 -0.5593  1.8262 -0.5972  0.6552  1.9813
#>  0.7490 -0.7762  0.0453  0.8314  0.3196 -1.6702  0.2315  1.7450 -0.5431  0.2478
#> -1.1093  0.2799 -0.3401 -0.7141  0.1733  0.0757 -0.9397  1.0321 -1.6319 -0.1988
#> -0.9192 -0.4878 -0.1895 -1.5815 -1.0083  0.3811  0.5888  0.0976  0.0746 -0.0329
#>  2.1358  2.9369 -1.2483  2.8356 -0.8351  1.1697 -0.1498 -0.4665 -0.1817  1.0095
#>  0.3370  1.8343  0.9244 -0.1673 -0.1796 -0.8828  1.4351  0.5655  0.3287  0.7705
#>  0.7037 -0.7053 -0.4998 -1.4785  0.6071 -0.7887  0.6887 -0.0758  0.6315  1.6298
#>  0.6335  0.0333  1.4335  0.6575 -0.6978  0.3713 -0.6069  0.7962 -0.5096 -2.1196
#>  0.0163 -1.4117  0.1764 -1.6096 -1.6724 -2.4144  0.8467 -0.5901  0.1783  1.2421
#>  0.9535  0.9789  1.1062  1.3680 -2.1456 -1.7783 -1.0448  0.6735 -0.9455  2.4079
#>  2.0568 -0.9221 -1.1709 -0.1735 -1.3820  0.9111 -2.2042  0.1466 -0.3567  0.6419
#> -0.9903 -0.6058 -0.6374  0.1882 -0.2789 -0.5651 -0.5036 -1.0962 -2.1553 -0.0580
#>  0.4367  1.6958 -0.3293 -2.7788 -0.5124 -1.5466 -1.0573 -0.3991 -0.6152 -0.0951
#> -0.6346  0.4272 -1.1016  0.8913  0.8124  0.0212 -1.6648  1.4339  0.7603  0.8470
#>  0.4500 -1.6705 -2.2586  0.4340  1.6129 -1.3413 -0.5219 -0.9923  1.0086 -0.4082
#>  0.9995 -0.3333 -0.3791 -0.8878 -0.0463  0.4134  1.4626 -0.2845  0.6889 -0.4735
#>  0.5544 -0.6694  0.9534  1.1222 -0.6800 -0.9830  0.1948 -0.4209  0.7349 -0.1637
#> -1.3772 -0.6544  0.4067 -0.7203 -0.0119  0.1066  0.2414  0.0362  0.9265  0.2828
#>  0.6448 -0.5928 -1.2138 -0.8615  0.4215 -1.6242 -0.0559 -0.0936 -1.2544  1.0254
#>  0.7591  0.7906 -0.5593  1.5707  0.1612 -0.1315 -1.0258  0.3094 -0.0868  1.2461
#> -0.8633 -0.5216  0.1090 -0.1925  1.0342  0.8464 -1.3347  0.2205  0.4388  0.6566
#>  0.8959 -0.0295 -1.7217 -0.9985 -0.7706 -0.5819  0.2219 -0.2576  1.0397  0.5890
#> -0.5375  1.2121 -0.8344 -0.0233  0.8298 -1.9864  0.2694  1.5921 -0.0199 -0.4741
#>  0.4751  1.4514 -0.6947  1.5608  0.2642  0.6720  0.4337  0.9013  1.6464  0.5208
#>  0.2540 -0.3813  0.5134 -0.2164  0.3218 -1.5483  0.0377  1.9330 -1.0394  0.4594
#>  1.2275 -0.5304  0.5164 -0.9517 -1.3227 -0.5505 -0.4186  1.2059  0.1591  0.6775
#> -0.4161 -0.1985  0.1782 -1.9926  1.1950 -0.0718  0.0786  0.0642 -0.7788 -0.7966
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]