Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.4808 -1.5125  1.6607 -0.0488 -0.8929  0.1339  2.8481  0.6215  0.1167 -1.5796
#> -0.0978  0.8553 -0.0395  0.3899 -1.7224 -1.8987  0.2586  0.8755  0.0026 -0.6030
#> -1.6204  0.1350 -0.8024 -0.6633  1.7706  0.0926 -1.9316  1.4752  0.9256 -0.8783
#>  0.5765 -0.7079  0.0059 -0.4421  1.7253 -0.1941  1.0842  0.6172  1.8267  1.0866
#> -1.1032  0.3622  1.4540 -0.6958 -0.7419 -0.6107 -0.2484 -1.5104 -1.0213 -0.9089
#> -0.9130  1.7876  0.2417 -0.8794 -1.4519  1.1253 -1.5398  0.3977 -1.0907 -0.5999
#> -0.1959 -0.6353  1.2387 -0.2465  0.1496  0.7901  1.1655 -0.7671  1.2257  1.9339
#> -0.7895  0.9237 -1.2969  1.7019 -0.3468  0.1656  0.3408  1.0755  0.8179  0.9790
#> -0.7170 -0.2773  0.5958 -0.7936 -0.4995  0.3034  0.1375  0.6958 -2.2942 -0.6610
#> -0.5838 -0.8641 -0.7857  1.4830 -0.2572 -0.9939  1.7943  0.2895  0.0846  0.1898
#> -0.2422 -0.5897  1.6431 -0.8747  0.2184  0.1169  0.4797  0.6488  0.3689 -0.3788
#> -1.0108  1.6658 -1.4253  2.0361  0.0126  0.0156  0.3342  0.1190  0.0267 -0.9772
#> -1.6019 -0.1215 -1.3119  0.1709  0.6505 -1.2314  0.9357  1.8921  0.8908 -2.5649
#> -0.0330 -1.0844 -1.7635  0.2602  0.1453  0.9908  0.0401  1.3430  1.0294  1.8585
#> -3.0566 -0.2837 -2.3601  2.2204 -0.1646  0.1808  1.2172 -0.6073  1.6360  0.0588
#>  1.2423  1.2649  0.1021  0.5538 -0.5878  1.0920 -0.1705  0.4503 -0.2484 -1.2107
#>  0.3120 -1.6849  0.5904  0.6928  0.0435 -1.5829 -0.4750 -0.0687  0.0397 -0.3206
#> -2.6259 -0.6496 -0.5487 -0.1380  0.0939 -1.2473  0.1572 -0.5088  0.7634  1.2492
#>  0.2614  1.1438 -0.9773 -0.4182 -1.4691 -1.6645 -0.4924 -0.2771 -0.0107  0.1315
#>  2.2122  1.0027  1.0142  1.0375  1.0552 -0.8530  1.3653  0.0574  0.2021  0.7805
#> -0.5876 -1.4838  1.0650  0.6154  0.1822 -0.7728  0.2309 -1.5512 -2.0452 -1.0383
#>  1.0837  0.7248 -0.1953  1.5967  0.9053 -1.1835 -1.1897  0.1992  1.6520 -2.5147
#> -1.8403 -0.3086 -0.6554 -0.2113  0.2135 -0.2595 -0.0242 -0.7799  1.1932 -1.0219
#> -0.5567  0.8076 -0.9632 -1.0605  1.3021 -0.9830  1.3940 -0.5210 -0.9367 -0.5624
#>  0.5769 -0.3501 -0.8992 -0.5128 -0.2368 -0.1426 -2.2603  0.0627 -0.6329 -2.0185
#>  2.3148 -0.1820  0.1410  0.5563  0.1528 -0.5679  0.2280 -0.9697 -1.0554 -0.3814
#>  1.5037  0.0431  0.7257  0.2184  0.1950  0.3980 -0.6981 -1.7121  0.0312  0.7380
#> -1.4007  0.7173  0.8387 -0.2536 -1.2530 -0.8450 -0.7680  0.5309 -0.7794 -0.3863
#>  0.8329 -0.9819 -1.0500  1.4781  1.1259  0.1631  1.5125 -1.0133  0.1095 -0.1503
#> -0.1330  0.9678  0.3992  1.4208 -1.6276  0.5186  1.0704 -0.9185  0.9104  1.5849
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]