Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.1995 -0.1721 -1.4050 -0.1016  1.1605 -0.0054  0.7649  0.7350 -0.2301 -0.0568
#>  0.6779 -1.8756  0.5861  0.4442 -0.4422  0.1594  0.8890  0.5713  1.6132  0.7505
#>  0.3883 -1.1585  1.0190 -0.7852 -0.3337  0.5787  0.7322  0.7820 -1.0221 -0.3346
#> -0.0579  0.9759 -0.1698  0.9235  0.1201  0.8556  0.8960 -0.7331 -2.1618  1.0960
#> -0.0459 -0.7122 -0.9536 -0.4516  1.4292  1.2388 -0.6551 -0.3339 -0.3778 -0.3474
#>  0.1801 -1.1155  0.6474 -0.1038 -1.1795 -0.1105 -1.8786  0.1647  0.4034 -0.1341
#> -1.4161 -0.3322 -0.2894  0.5294 -0.0679 -0.2471  0.9758 -0.0260  1.4961  0.5807
#> -0.7160 -0.2581 -1.2469 -1.1079 -0.3585 -0.8783  0.4399 -0.3622  0.6229 -1.9491
#>  0.7140  0.7753 -1.5250 -1.8335 -0.4143  1.2008  1.6856 -1.4269 -0.2074 -0.3692
#> -0.9019  0.7422  1.2024  0.4150  1.0675 -1.0169 -0.8848  0.7940 -0.7166 -1.3860
#> -0.0532 -0.4582  0.3842  0.1201 -0.5681  2.0344 -1.0992  0.4149 -1.2809  1.7552
#> -0.7991 -0.9211 -0.0321  1.8372  1.9177 -0.3384 -2.3445  0.2091 -2.0739 -0.6298
#>  0.4305  1.5236 -0.7439  1.7115 -0.7161  0.5835 -0.4190  0.7161 -0.0598  0.4761
#>  1.1430 -0.2915 -1.3393 -0.9084 -0.5453  0.7217 -1.3454 -0.6006  0.8901 -0.0178
#>  0.8326  0.3141 -0.6120 -0.4556 -0.4530  0.3282 -0.4602  0.2168 -0.7511  0.8234
#>  0.0203  1.1919  0.1945  2.3152 -0.6310 -2.0295  1.9405  0.1766  1.3958 -0.7401
#> -2.3031  0.4057  1.2836  2.0533  0.6941  0.0891  1.1556  1.7860  0.1006  0.4783
#>  0.3179 -1.6370 -0.7492  0.9818  0.8487 -0.1694 -0.4526  0.4444 -1.0150  0.8788
#> -1.5139  0.3222 -2.0011  0.8085  0.6995 -0.2257 -0.9820  2.0697 -0.1548 -2.2648
#> -1.1223 -0.6329 -1.2745 -2.4748 -0.1824 -0.5667  1.4216 -0.3333 -0.5865 -0.7845
#>  1.3130  0.6744  0.5075  1.3159 -0.6231  0.4310  0.5201 -0.8577 -0.9434  1.5104
#> -0.4711  1.5343  1.7385  0.2411 -0.1466  1.6413  0.7672 -0.0076 -0.8271  0.7659
#>  1.5956 -1.1671  0.5768 -0.4599  1.6911 -1.4869  0.0938  0.7259  0.4261 -1.0351
#>  0.1784  0.2868 -0.3629  0.7123 -0.2844 -0.3229  1.5493  1.4080 -1.0388 -0.6309
#> -0.5429 -1.2555 -0.3736  0.3752 -0.7848 -1.4922  1.6563  0.2577  0.3507 -0.1742
#> -0.8067 -0.1596 -0.5556 -0.7127 -0.2638  0.6745  0.3701  0.6477  0.7954  0.5402
#> -0.6759  0.4008  0.4579  0.0941 -0.6417  0.8498  0.2548  1.5772  0.7202 -2.2100
#> -0.5411 -2.6265  0.8064  0.7122 -0.8916 -0.0298  0.3194  0.6761  0.3380  0.9730
#> -0.1304 -0.2695  2.0300  2.9249 -0.2625 -0.8410 -1.5225 -0.6666 -0.7452 -0.7209
#>  1.7498 -0.9651 -0.3944 -1.8852 -0.4986  0.1326 -0.4390 -0.7834 -0.1420  0.0055
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]