Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4649  1.9895  1.4541 -0.3336  0.1175 -1.1412  0.2619  0.0962  1.2659 -1.6037
#> -0.5625  1.3181  0.3807  0.4539  1.3397  0.7335 -1.9837 -1.1657  1.3788  0.1652
#> -0.1964 -0.8970 -2.3718 -1.5363  0.5422  0.4027  0.3183  0.9491  1.4556  1.3413
#>  0.0837 -1.1926  0.4065 -0.4546 -0.6559  0.6375  0.8142 -0.3970  0.3330  2.3326
#>  0.2889  1.7637 -0.1914  0.3365 -0.0627  0.2765  0.8261 -0.0074 -0.5310  0.0844
#>  1.1808  0.1047 -0.7270 -0.7811  1.1919  1.9098  1.4111  1.2080 -0.3307  1.0671
#>  0.3394  0.6069  0.7131  0.2206  1.0733  1.4924 -0.5253 -0.0807 -2.1402  1.7398
#> -0.0314 -1.2806  0.4419  1.0807  0.7303  1.4317  0.0559  1.0627  0.3085 -0.9078
#>  0.7273  0.6197  0.6376  0.0304  2.2234 -0.3542 -0.8521 -0.1146  1.0575 -1.3276
#> -0.4053 -1.7297 -0.8342  1.2395 -1.8258  0.5533  0.8711  0.9884 -1.2844 -1.8415
#>  0.1925  0.2323 -1.4441 -0.6919 -0.1304  0.2867 -0.6835  0.0490 -0.4019 -0.9026
#> -1.0043  1.6467  1.2943  0.0090 -0.8394  1.7996 -2.7556 -1.6807  1.4219 -1.8498
#> -0.2330 -0.5930 -0.2416  1.5337  1.5626 -0.8968 -1.1299  0.3942  0.2074  0.2904
#> -0.0840  1.8229 -2.9940  0.5700  0.3897 -0.7965 -1.4963 -0.6086  0.2405 -1.1904
#> -0.1651  0.3382  0.0262 -1.4382  0.5954  0.6800 -2.0234 -0.5633  0.5356  2.9800
#>  1.8236 -1.3193  0.8631 -0.6930  2.0660  0.7042  0.5518 -0.0251 -0.6697 -0.6897
#>  0.0470 -1.2319 -0.4606  0.5881 -0.4523 -2.5500 -1.4299  0.4974  1.9138 -0.9732
#> -0.5733  0.9263 -0.2376 -1.7145 -0.1489 -1.4223 -0.9379  0.0645 -0.2425 -0.9127
#>  0.8566  0.0354 -0.5173  0.7372 -0.0765 -0.4179 -0.0248  0.6565 -1.4783  0.1777
#>  1.0052  1.1261  0.6819 -1.6241  1.1771 -0.4109  0.5012 -1.4101 -0.1972 -1.1152
#> -0.9681 -2.0352  1.8647  0.5286 -1.3546  0.9887 -0.0357 -0.7164 -1.2549 -0.5740
#>  0.3449 -1.0144  1.5519 -0.1093 -0.1113 -0.6605 -0.5837  1.7953 -0.8253  0.1622
#> -0.2054  1.1866 -0.8111 -2.2011 -0.5319  0.3370 -1.8422 -1.7426 -1.2987  0.2607
#> -0.9598 -0.8953 -0.1485  0.2195 -0.3338  0.3564  0.5706  1.2455  0.2242  1.1458
#> -0.2549 -0.7591  0.0538  0.4933 -1.4179  0.0018  0.1204 -0.0619  0.4318  1.4471
#> -0.6932 -0.6504 -1.2657 -1.6461  0.4038 -2.2695  0.4900  1.3798  0.5902 -1.5380
#> -0.3948 -0.1082  0.1345  0.5097  0.3446  0.5229 -0.1964  0.2613  0.4225 -0.3512
#> -0.9205 -0.3671 -0.9774  0.6264  1.7999  0.3064 -1.8853  0.7725 -2.4921 -0.6937
#>  1.1622 -1.2603  0.1515  1.8088 -0.2913  0.8851  0.9066 -0.4680  0.9478  1.7153
#> -2.4258  1.2778  1.5948  0.4585  0.1110 -1.7938 -0.7803  0.8691 -0.2926 -0.6080
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]