Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0097 -0.5794  0.1717 -0.9382  1.2192 -1.1779  1.2260 -1.5095 -2.3444 -0.6393
#> -1.2216 -1.9793  1.0969  0.9384 -1.1997  0.3144  0.9230  0.9807  0.6314  0.1683
#>  1.6316 -0.2295 -1.2231  1.0636  0.2268  1.6818 -0.3592  1.0442  0.9661 -0.7903
#>  0.3968 -0.2098 -0.8206  0.3630  0.2140  0.0623 -0.3278 -1.4494 -0.3559  0.1615
#> -1.3365  0.6215  2.0724 -0.5877  0.8521  0.5296 -0.9479 -1.2618 -1.3692 -0.0894
#> -0.5844 -1.4331 -1.2937 -1.1303  0.4856  0.3305  0.2821  1.5428 -1.2395  0.2141
#> -1.8313 -0.7504 -0.6052  0.7607  0.0271  2.1981 -0.0570 -0.9030 -1.2260  1.5149
#>  0.9600  0.7665 -1.1031 -0.6122 -0.9290  0.7125 -0.3543  0.3844  1.7635  1.5506
#>  0.6889  0.0816 -0.5487  0.8593 -0.2925  0.6815 -1.0116 -0.7954  0.3562 -0.1492
#> -0.6332 -0.6582  0.3179 -0.7680  0.6644  0.5467  0.9598 -0.3229  0.4139  1.3811
#>  0.0217  0.5010  1.5083  0.9199 -0.1991  0.3212 -1.1157 -0.5944  0.0533 -0.1632
#> -0.4744  1.9013 -0.7990  0.2112 -0.8026  0.4627 -0.4568  1.9114 -0.2134  0.1153
#> -0.0870  1.5905 -0.3012  0.5941  0.1135 -0.3956  0.1932  0.8590 -0.2645 -0.7378
#>  0.0947  1.7954 -0.2425 -1.0176  0.1923  1.2032  0.3335 -0.3741  0.7022  0.9264
#>  0.2246 -0.8207 -1.7481  1.1813  0.7708 -2.7967  0.2312 -1.3831 -0.5052  1.5567
#> -0.0457 -0.8795  0.4031  1.2217  0.0426 -0.6437 -0.2524  1.1377 -0.8155  0.5844
#>  0.4746 -2.3916 -0.7540  0.2762  0.0370 -0.6213  0.8574 -0.6150 -0.3248  0.2287
#> -0.0475  0.3721  0.4883 -0.5266  0.9683 -0.5122 -2.4471 -1.0080 -0.1860 -0.2032
#> -1.2115  0.1302 -0.0631  1.6229 -1.2460 -0.1959  1.4588 -0.0439 -1.6619  1.2053
#>  0.4747  0.7978  0.4217  0.3949 -0.4826  0.6883  0.5307  0.0218  0.7954 -0.8281
#> -0.3148 -1.5572  0.6880  0.8113  0.1576 -0.9636  0.1856 -0.8966 -1.2913 -0.2099
#> -2.2044 -1.6004 -0.2840 -0.3198  0.7566  0.2727 -0.0944  0.9960  2.2597  0.7177
#>  0.3586 -1.8484 -0.5440  1.0889  0.5894 -1.8423 -0.9021  0.0806  1.5469  0.2395
#> -1.2864 -2.6060 -0.8221 -2.5992  0.1451  1.0091 -2.1431  0.3858 -0.2187  0.9950
#> -0.9800 -0.2041 -0.1772  0.6740 -0.5377 -0.3719 -1.0474 -0.4399 -0.1231 -0.7919
#> -0.3799 -0.1098  0.7015  0.1313 -0.7531  0.1784 -0.3351  1.3145  0.0710  0.4281
#>  1.3918 -0.3456 -0.5030  1.1313  0.8266  0.3798  3.2778 -1.1538 -0.6897  0.1772
#>  1.0660  1.6048  0.8349 -2.4274  1.4293 -0.1683  1.9239  0.7521 -1.2328 -1.3992
#> -2.5797  0.4361  0.3320  1.2236 -0.6602  1.1230  0.0851  0.8583  0.7844 -0.6923
#> -0.2278 -0.2084 -0.2244  0.4986  0.7362 -0.2849  0.1497  0.2955 -0.2974 -1.0369
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]