Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 1.2827e+00  7.8198e-01 -9.4136e-01 -7.7633e-01  1.6141e-01 -1.2584e+00
#> -6.9302e-01 -1.1350e-01 -1.7552e-01  3.0424e-01  1.6404e+00 -2.2520e-01
#> -6.6871e-02 -3.8273e-01  1.5600e-01 -5.9729e-01  4.5709e-01  3.9881e-01
#>  1.7681e-01 -3.0638e-01 -2.7291e-01  1.4843e-01  7.3236e-01 -1.2732e+00
#> -6.7067e-01 -1.1105e+00  1.0731e+00  1.7152e+00 -5.8746e-01  8.8968e-01
#>  7.6553e-01  5.7075e-01  6.0497e-01  1.7675e-01 -7.4435e-01 -3.6609e-01
#>  1.6985e+00  1.3165e+00 -4.8928e-01 -1.5526e+00 -1.5098e+00  2.1308e-01
#>  1.6626e+00  8.0167e-01  7.1301e-01 -1.1940e+00  2.2002e+00 -8.7586e-01
#>  5.6180e-02  8.4700e-01 -7.8945e-01 -7.1875e-01 -1.2434e+00 -5.0843e-01
#>  6.9527e-02  1.5676e+00  2.3044e+00  1.7352e+00  8.4202e-01  1.0199e+00
#>  1.6597e+00  1.0012e+00 -5.2924e-01  6.5606e-01  7.2917e-01 -5.1969e-01
#> -6.5754e-01  1.2139e+00  7.5455e-01 -1.6038e+00 -8.6733e-01  4.0053e-01
#>  9.3409e-01 -1.2594e+00 -3.8582e-01 -1.5538e+00 -7.4713e-01  5.5422e-01
#> -3.7566e-02 -4.6934e-01 -8.2748e-01 -8.7333e-01 -6.6344e-02 -1.4672e+00
#> -8.1504e-01  3.9098e-01  1.2000e+00  1.1492e+00  5.5172e-02 -7.0683e-01
#> -1.5524e-01  1.9795e-01 -1.2765e-01  1.0416e+00 -6.7499e-01 -4.1979e-01
#> -1.6488e-01  2.6492e-01 -4.2649e-02 -6.2912e-01  3.8906e-01  4.2580e-01
#>  9.9602e-01  2.4591e-02  2.2515e-01 -2.0290e-02  7.3139e-01  3.1523e-01
#> -1.9249e+00 -1.6668e-01  1.3010e-01  6.4477e-01  8.9148e-02  4.9650e-01
#>  4.2874e-01  5.7924e-01 -7.1379e-01  2.5357e+00 -1.5798e-01 -2.0566e+00
#> -4.1512e-01  8.6961e-01  2.8703e-01  4.3487e-01  5.8078e-01  6.3335e-02
#>  2.4416e-02  9.8236e-02 -6.4906e-01  1.4145e+00 -4.6431e-01  1.2457e+00
#> -7.0578e-01  1.7747e-01 -3.6314e-01 -2.7763e-01  8.8643e-01 -3.0011e-01
#> -6.6042e-01  3.2595e-02 -5.9417e-01  3.9378e-01  6.0015e-01 -2.0729e-01
#> -2.7598e-01 -2.5489e-01 -6.4819e-01 -3.3051e-01  9.2977e-01  1.3678e-01
#>  6.8230e-01 -7.4137e-01 -5.8556e-01 -3.8007e-01 -3.9500e-01  5.3199e-01
#>  1.6803e-01 -5.1717e-01  2.2269e+00 -3.1928e-01 -1.1534e+00  1.7343e+00
#> -5.8537e-01 -9.1030e-01  2.9649e-01 -6.6607e-01  8.0867e-01 -1.1261e+00
#>  1.2683e+00  7.2838e-01  4.4760e-01  3.1873e-01  9.6198e-01 -7.7922e-01
#> -3.3234e-01  5.7812e-01 -4.3794e-01 -1.5070e+00  7.5482e-01 -6.6592e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]