Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.1262 -0.1787  0.1416 -0.6049 -0.8570  0.4665 -0.8240 -0.3505  0.1631  0.4980
#>  1.1389 -1.0019  0.4107 -1.3134 -1.7815  1.4766  0.2183  0.0164  1.1223 -0.7881
#>  1.9469 -1.1086 -0.2588 -0.0455 -1.7974 -0.0340 -0.6627  1.3713  1.0582  1.4256
#>  0.6318  0.3249 -1.8685  0.8545  0.8754 -2.2127 -0.3735 -0.8905  0.8073  1.6047
#>  1.8005  0.3847  0.6632  0.7911  0.8055 -0.4251 -0.1485  1.4193 -0.8917  0.3271
#>  0.1733 -1.3166 -0.9929  0.0031  0.3502  0.4861  0.1809 -1.3619  0.6967  0.4853
#>  0.5886  0.8018  0.5046 -0.6845 -1.7067 -0.7019  0.1665  0.6211 -0.9006  0.1168
#>  1.1138  1.4665 -0.7999 -0.4478 -2.2200 -1.1480  0.1615 -1.1727  1.5974 -1.4548
#>  0.4858 -1.0024 -0.2162  0.8961 -1.4941  0.1453  0.2978  0.6051 -0.2342 -0.1303
#> -0.7747 -0.8262 -0.2546 -0.8804 -0.5158 -0.0732  1.1221  0.8342  0.2174 -0.9173
#> -0.1905  0.1278  0.5127 -0.7667 -1.1130  0.0689 -2.2212 -1.3014  1.4944 -1.0192
#>  0.5383  1.0843 -1.1626  0.9809 -1.2792  0.0572  0.0006  0.7821 -0.8261  2.8920
#>  0.6160  0.4321 -1.7685  0.2878  1.2894  0.0381  0.0684  0.1338 -0.1577  1.4229
#> -1.2194 -0.5912  1.0719 -0.3174  1.1009 -0.5440  1.7483  0.2462  1.6844 -2.0040
#> -1.1448  0.4189  0.1214  1.9481 -0.7848  0.7492 -0.1835  0.4154  1.4926 -0.5955
#> -0.5189  0.6134 -0.1555 -1.0980 -0.4091 -0.5436 -1.0981  0.9291 -0.2537  0.9698
#>  0.3795 -1.8116  1.0319 -1.3527  0.4182  0.2055  0.8975 -1.0410  1.0342  1.0055
#> -1.3714  0.5730 -0.0324 -1.5463 -0.0120  0.0630  0.3667 -0.9909  0.8509  0.4582
#>  1.1918 -0.5696 -0.9733  0.4558  0.3805 -0.4536  0.7331  1.3541  2.8758 -0.8012
#>  0.8132  1.6458  0.0695 -1.3847  0.8786 -0.8867  0.2024  1.1263 -0.5665 -0.3928
#> -0.9878 -1.0229 -1.0198  0.5417  0.4129  0.1096  0.0729  0.9663 -1.6039 -1.3846
#> -1.2007 -0.3101  0.3927 -1.3785  0.2252  0.9597  0.1251  0.0722  1.1374 -1.0531
#>  1.1550  2.4321  1.6638  0.9003  0.5897  0.4870  0.5872 -0.2055 -1.2439 -1.0656
#>  0.2571 -0.3273  1.3010 -0.9330  0.5956 -0.9593 -0.6107 -1.8365  1.3946  2.7729
#>  0.6055  0.9239  0.4489  0.4725 -0.5430  0.5898 -0.4111  0.5425 -0.4056  1.0618
#> -1.3180 -0.3015  0.0065 -0.5182  0.5265 -0.9405  0.1348  0.1820 -0.7148  0.5951
#> -0.0591  0.2432 -1.0926  0.6623  1.8373 -1.9196  1.3080 -0.4136 -1.3082 -0.1737
#> -0.5216 -1.0572 -1.6245  0.3171 -1.2300 -0.9176  0.7333 -3.1399  1.5047  0.0816
#> -0.2285 -0.8585  1.5934 -0.8811 -0.2088 -1.1868  0.8534 -1.2666  0.1812 -0.9358
#>  0.6484 -0.4014 -2.1943  0.0311 -0.7572  0.2169  1.2221  1.0421 -0.8929  1.1751
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]