Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.1850 -1.5037  0.4443 -0.9723  0.5877 -1.7060  2.6350  0.4948 -0.8299 -1.6141
#>  0.1138  2.6249  0.7788 -1.4063  0.4124  0.2362  0.6607  1.6412 -0.1419  0.3999
#> -0.0754  1.2875 -0.4590  1.7424  0.0655  0.3543  1.3301 -0.3435 -0.5313  0.3134
#>  0.2834 -0.5929  0.0172 -0.4517  1.7209  1.3536 -0.1719 -0.6376  0.1150 -1.4974
#>  0.9428  1.8997  0.0516 -0.8961  0.3595 -1.2514 -0.3225  0.5615 -0.5695 -0.7597
#>  0.5606 -0.1792 -0.4837  0.5185 -0.9219  0.5619  0.6415  1.2094 -0.9638 -0.7685
#>  1.2322 -0.7774 -1.9707 -0.1877  0.9883  0.1926  1.6301  0.0648  1.6065  0.7524
#> -0.6524  0.2442  0.2249 -0.5413  0.3852  0.2770 -0.0316  2.5680 -0.7402 -1.2777
#> -0.4686 -0.5999 -0.4320  1.0094 -0.3192  0.7929 -1.0526 -1.5138 -0.2696  0.2486
#> -0.6953  0.3953  0.8056  2.0247  0.1913  0.2565  0.6894  1.2684  1.2117  1.3914
#>  1.1867  0.9477 -0.4504 -2.1597 -1.7537  0.5790  0.4787  1.5990  0.3195 -0.9559
#>  0.5001  0.6854 -0.2295  1.1529 -0.5184  0.7020  2.4300 -2.5742  0.3970 -2.3321
#>  1.8071  0.0979 -0.3983  0.5009 -0.2257  1.0001 -0.0249 -1.2132  0.3796 -0.0651
#> -0.2191  0.6603  0.6372  0.9324  0.4549  0.5492  0.8574  1.0084  0.6992 -1.2246
#> -0.4384 -0.3255  0.2120  1.1090 -0.0507  1.7095 -0.4645 -1.6297 -0.0821  0.0165
#>  1.2800  0.5403 -2.4731  0.5375  0.6024  1.5839  0.2141 -0.3998 -0.1721  1.0623
#> -0.9520  0.2063  0.6405  0.0570  0.2237  0.5538  0.3463  0.0211 -0.3735 -0.4005
#>  0.3113  2.0475  0.8239 -0.4633  0.0432 -1.5995  1.1833  1.1454 -1.3078  0.2846
#>  1.6074 -0.5285 -0.3052 -1.0096 -0.1933  1.2807  0.8383 -0.1986 -0.9150 -1.3266
#>  0.0498  0.1522 -0.4777  1.2009  0.8392  0.1785  2.2632  1.5526  1.0679 -0.3102
#> -1.5394 -0.6425  1.5550 -1.1662 -0.6468  1.5852 -0.9369 -1.8428  1.9006  0.4576
#> -0.3358  0.2236 -0.3393 -0.7122  2.1685 -0.1781 -0.2409  0.7597 -0.5462 -1.0734
#>  0.4465  0.6560 -1.1772  2.1772  1.0095 -0.2877 -1.0785 -1.8119  0.3237 -1.6775
#> -0.0059 -1.5519 -1.4116 -0.4133  0.0762 -0.3908  0.3671  0.5291 -2.4964  0.4742
#>  0.6340 -0.9347 -1.1502 -0.9507 -0.0797  1.2466  0.0050  0.4971  0.4016  0.9071
#>  0.0808  0.2017  1.3533 -1.0310 -1.9543  0.5658  0.9537  0.9372  1.3995  0.2564
#> -0.9724  0.2168  0.5392  1.0298  0.8326  0.6450 -1.1356  0.8808 -1.1920 -1.4598
#>  0.8817  0.5724  0.3386 -0.8561 -1.1549  0.1669 -0.0673 -0.8063 -0.5392 -1.5907
#> -0.0310 -0.0245 -0.3786  0.8381  2.0386 -0.3331 -0.7846 -1.4090  2.0582 -0.9724
#> -0.9299 -1.1475 -1.4747 -0.8596 -0.2833 -0.6065  0.9221  0.7773  0.5565  1.2470
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]