Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.1660 -0.8365  0.7340 -0.4912 -1.2057  2.5303  0.9920  0.6880 -1.9233 -0.9538
#>  0.6895 -0.4104 -2.1816 -1.1322  1.0330 -1.9317 -0.9213  0.1838 -0.8324 -2.1591
#> -0.8608  0.7075  1.4276  0.7308 -0.4582 -0.7707  1.6942  0.6330  0.9009  0.1461
#>  0.3389 -1.1047  1.1190 -0.6567 -0.4764 -0.0845 -0.2205  0.7322 -1.0726 -1.2859
#>  0.3970 -0.1033 -0.9521  0.9340 -1.7248 -1.4033  1.2224 -1.7886 -0.8809 -0.9686
#> -0.3349  0.3006  0.0617  0.1847  0.1083  0.5457  0.6090 -0.4158  1.7617  1.0851
#> -0.2616  0.4552  1.1273  0.8745 -1.0613  1.1952  0.1368 -1.1402  0.2598  0.1486
#>  0.5691 -1.3742 -1.4985  0.6358 -0.5412  0.6331  0.0882 -1.7497 -0.6304  0.2571
#> -0.4199 -1.1820 -0.1777 -0.4269 -1.5013 -0.1703  0.6574  1.8150  0.7051 -1.0342
#>  0.8511 -1.6452  1.0236 -0.1010 -0.8543  1.3606 -1.6411 -0.5350 -0.4902  0.6535
#> -1.1629  1.0108 -0.0907  0.3754  0.0394 -0.4303 -0.7841 -1.1280  0.5107  0.6071
#>  0.0036  0.6472 -0.7543 -0.4542  1.3182 -2.9791  0.8215  0.3286 -0.9338  0.4291
#> -0.2638  1.6956  0.2667 -0.5282  0.0729  0.2946  0.3831  0.1584 -0.9770 -0.4240
#> -0.2159  2.2619 -1.4236  0.3905  0.4860  2.6670  0.5794 -0.7412 -1.0107  0.8327
#> -0.5580 -1.1278  0.8156 -0.9245 -1.2456 -1.4741  0.6755 -0.4163  0.8227  0.1994
#>  0.8307 -0.7051  0.2977 -0.8595  2.1096  1.6034  0.3080 -0.0576 -1.3060  1.8421
#> -0.4433  0.0128 -2.8918 -0.3432  0.1948 -0.6792  1.3649  1.6958  0.8722  0.4634
#> -0.2552  0.3789 -1.1761 -0.3390 -1.4697  0.1276  1.7498  1.5902  1.4106  1.5532
#> -0.9473  0.6534 -1.9574 -1.3158  0.6884 -1.2894  0.0067  0.3550  1.9806 -0.8535
#> -1.3234 -1.7735 -1.1267 -1.9469  1.3194  0.0957 -0.9386 -0.5859  0.6412  1.5760
#>  0.3949  0.3762  0.6972 -1.0770  0.5559 -0.0118 -1.2664  0.4826 -1.6003 -0.6374
#> -0.7213 -0.5564 -1.6402 -0.0017  2.1280 -1.3671  0.3243 -1.1260 -1.0844 -1.0354
#> -1.2850 -0.1812 -0.8509 -0.1931  1.1808 -1.4288 -1.4831  1.1240 -0.4873  0.0471
#>  0.6041  0.9184 -0.4643  0.3339  1.2421  1.1637 -0.6586 -1.1986 -0.1882  0.4275
#>  1.2922 -1.5749  0.4309  0.6202 -1.3547 -2.4958 -1.3514  0.4968  0.8910  1.2838
#> -0.3031 -2.0198  1.8355  0.6030  0.3280 -1.1601  0.4219 -1.3212  1.1350 -1.1536
#> -1.5914  0.2051  1.4844  0.5916 -0.5053  0.6551  1.6102 -2.4593 -0.2229  0.3988
#> -0.0129  0.0907 -0.1961  0.8598  0.0639  0.0586  0.4642  0.4721 -1.0407 -0.5555
#> -1.4702 -0.3756 -1.0475 -0.1668 -0.0128 -0.1935  1.3479  0.0866 -2.2023  1.0977
#>  1.2401  0.1597  0.1819 -1.2567 -0.7713  0.0825  0.2596 -0.4414  0.1625 -0.6712
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]