Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.5367  1.9382 -1.7021 -0.7943 -0.3066 -0.6263  0.3606  1.5179  0.5761 -0.1098
#>  1.0909 -1.8067  0.7405  1.5387 -1.4297 -0.0204  0.8839 -0.8043 -1.3241 -0.0356
#> -1.0577 -0.3958 -0.0266  0.5935 -0.5397 -0.7651  0.5137 -1.9248  1.6806 -0.0591
#>  0.3045  0.5251 -0.3340 -1.7330 -1.2069  0.8922  0.6939  1.2579  0.3847 -0.7221
#>  1.1029  1.4617 -0.7679  1.0461  0.5359 -0.6266 -0.7756  1.6026  0.9523  1.1266
#>  1.5325  0.8177  2.3153  1.6931 -0.4227 -1.6227  1.1295 -0.1777  0.8179 -0.8391
#>  0.3563  0.3331  1.2890 -1.3541 -0.1725  0.8052  0.2299  0.1056 -1.3679  1.7649
#> -1.0756 -0.5413  0.1381 -0.2229 -0.9341 -2.2292  0.7880 -0.3615  0.0051 -2.8245
#>  0.0735  0.1812 -0.2460  1.4526 -0.9324 -0.8147  0.8853  0.6212  0.2236 -0.5298
#>  0.0221  1.6534 -1.3868  0.2727 -0.4669 -0.7631 -1.0879  0.2857  1.2136 -0.2811
#> -0.2628 -1.1235 -0.1993 -0.1645 -0.1113  1.3447 -0.8489  0.8717  1.6184  1.1535
#>  1.6011  1.1263  0.2828  0.5760  0.7194  0.3097 -0.0658 -0.7951 -0.7986 -1.1532
#>  0.8386 -1.0064  0.1636 -0.9545 -0.6023  0.3207  1.0481 -1.6622  0.6335  0.7415
#>  0.5349 -0.0242 -0.6424  1.4571  0.6326 -0.0300  1.3664  1.9976  1.2528  0.2785
#> -0.8653 -1.2761 -0.4658  1.2200  1.6931 -1.0255 -1.0894 -0.2007 -0.0783 -0.3944
#> -0.0940 -0.1769 -0.2522 -0.7625  0.9238  0.9134  0.1407 -1.2327  0.0596 -1.0655
#> -0.4206  1.9352  0.3705 -0.9768  0.0744 -0.3604 -2.5533 -0.8724  2.1671 -0.8584
#> -0.6640 -1.4886  0.4143  2.0231  0.9508 -0.1676  0.0456  0.5327  0.0117  1.0824
#>  0.6079 -0.4618 -0.6170  0.7631  0.2258 -2.0450 -0.5581 -1.9897 -0.5194 -1.3278
#> -0.0436 -0.4485 -1.2156  2.5465 -0.3278  0.2197 -0.0287 -0.8579  0.6566 -0.9726
#> -1.1633  0.9892  0.7087  0.5803  0.5864  0.1017 -1.1861 -0.5397 -0.1323 -0.1229
#> -0.2157 -1.5149 -0.3882 -1.3368  0.1659  0.2283 -0.0966  0.5478 -0.1453 -0.0710
#> -2.0360  0.1993 -0.4767 -2.1599 -2.0441  0.3141  0.5311 -1.5216 -0.4745  0.5066
#> -1.1283 -0.7276  0.3931 -0.7490  1.3694  1.4840  0.8430 -0.5079  1.9975  0.4338
#>  1.0058  1.7510 -0.5761  0.6608 -0.9394  0.1884  0.3253 -0.1912 -2.0644 -1.1398
#> -0.6886 -1.9601  1.6006 -1.3368 -2.2875 -1.0192  1.6806 -0.4681 -1.4127 -1.0200
#>  2.5180  0.4045  0.5806 -0.0125 -0.4872  0.8307  0.4414 -0.3780 -0.3835  0.9714
#> -0.2502 -0.5332  0.1210 -0.4324 -2.4486  1.2678 -1.2950  1.0024  1.7927  2.1005
#> -0.5247 -2.1775 -0.3184 -2.6386  0.9271 -0.3871 -0.5240 -0.3617  0.3488  1.2654
#> -1.3811  0.5276 -0.5442  0.4069 -0.7542 -0.1887  1.0688  0.9176 -0.5486  0.1649
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]