Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.3453  0.4445  0.3329 -1.1463 -0.2252  0.4637 -0.9641  0.3258 -0.1464  0.3349
#>  0.5347  3.3986 -1.6596 -1.8444  1.2220 -2.0001 -0.7194 -0.0101  0.1097 -0.9073
#> -1.0471  1.6486  0.3554  0.0129 -0.6481 -0.7329 -0.5298  0.2545  0.1438 -0.8035
#> -0.4603 -0.9536 -0.4229  0.3624 -0.9257 -0.3065 -2.4646  0.8252 -0.2282 -1.0308
#> -1.6976  0.1389 -0.1609  1.3515  0.0027  0.1517 -1.5873  0.5260  0.2067  0.8126
#> -1.6185  0.9469 -0.3892  2.2275 -1.1765  0.4893  0.9087 -0.6410  0.3354  0.6180
#>  1.2937  0.2466  0.2771  1.6804 -0.0875  0.5860  0.1728  0.3976 -1.1791 -0.3815
#>  0.9471 -0.5003 -0.0462 -0.3129 -0.3931  0.1676  0.3509 -0.6379  0.0985  1.6996
#> -0.4128 -0.5416 -0.5094 -0.8386  1.8099 -0.3498  0.8661  0.4859 -0.2236 -0.8112
#> -0.7620  1.2437  0.5105  0.4845  0.0565 -0.2072 -0.8774  0.8391 -0.5310  1.1232
#>  0.8933 -1.3689  2.4248  0.0405  1.2369  0.6140 -0.1036 -0.3803  1.8777  0.8290
#> -0.3407 -0.3934 -0.1835  0.8226  0.2030 -1.0211 -1.0505 -1.6404 -1.0203 -1.2390
#> -0.7460  1.0188  0.9730  0.1908  0.3812  0.3239  0.1348  0.4168  0.3532  0.1805
#>  0.1893 -0.3325 -0.5208  0.0284  1.0170  1.0660 -2.1392 -0.5500 -2.3930 -0.4597
#>  0.9132  0.5004 -0.0477  0.1030  0.3621 -0.5544  0.8299 -1.2297  1.1087 -0.1027
#>  0.8611 -1.5949  0.5980 -0.5138 -2.0787 -0.7155  0.5048  1.2617 -0.5995 -1.5642
#>  0.1700  0.2865 -0.0233  0.8294  1.1613 -0.4420  0.5715 -1.6098 -0.0937  0.9146
#>  0.2439  1.1045 -1.4708  1.2394  0.6572  0.0124 -1.5935  1.2131  0.0244 -1.0623
#>  1.3945  0.1681  0.5559  0.7071  0.2987  1.8632  1.4751 -0.1722  1.7229  0.4881
#>  0.7607 -0.4018  0.3193  0.4451 -0.8889 -0.9964 -0.6499  0.1306  1.4990  0.9931
#> -1.1237 -0.2835 -1.5060 -1.3472  0.9403  0.1005  0.9061 -0.4345 -0.3840 -0.2416
#> -0.5603  0.4920  0.0341  0.3242 -1.1415  0.8290 -0.1859 -0.5162 -0.7775 -1.0345
#>  0.0285 -0.1179 -0.0039  0.9091 -0.3021  0.7712  1.4703 -0.0100 -0.5855 -0.5336
#>  2.0698  0.8236 -0.6682 -0.2963  2.9703 -2.1341 -1.7676 -0.9690 -0.7139 -1.3643
#>  0.2286  0.2615 -1.1947 -0.5899 -1.0095  0.7665 -0.2916 -0.7465 -0.9667  2.2900
#>  0.7346 -0.3632 -0.0301  0.5842  0.6043  0.5660  0.8382 -1.0688  0.7498  0.4011
#> -0.6134 -1.3562  0.0997 -0.4267  0.1996  2.3906  1.0548 -0.2604 -0.3276 -1.4921
#> -0.5818  0.5684  0.3915  0.6358 -0.3998  0.3970  0.5100  0.0276  1.3440  0.9907
#> -0.1439 -1.5669  0.5077  1.3929 -1.3986  1.0813 -1.6310 -1.1222 -0.9292 -1.1102
#>  1.8800 -1.4425 -0.3132 -0.9983  0.0175 -2.3563 -0.1768 -0.3030  2.2345 -0.0629
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]