Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.0377  0.6844  0.2813  0.1597  1.2109  0.6321  1.8187  1.9477  0.6217 -0.3885
#>  0.9755 -0.5426  0.1455  0.0784  0.1742  0.3200  0.0190  0.0027 -1.8546  0.8089
#>  0.6695 -0.1392 -1.2016  0.8386  0.4965  0.2600 -0.4378 -0.8684  1.1842 -0.2808
#> -0.0961  1.9398  2.2262 -0.7969  0.7219 -0.2860  1.5847 -0.1549  0.9573 -0.4453
#> -1.1088  1.6372  0.6640 -0.1904  0.4032 -1.8348 -1.1915 -0.6108 -0.5102  0.6204
#>  1.3915 -0.1441 -0.9384  1.0214  1.3763 -0.1880 -1.1550  0.9857  1.1101 -0.2492
#>  0.5500 -0.3933  0.7038 -0.0845 -0.1322  0.8445  0.2207  0.8250  0.0556 -0.5780
#>  0.1930 -0.9876  0.8829 -0.0459 -2.3676  1.2819 -1.4591 -1.7343 -2.6418  1.7115
#> -0.3436  1.0518  0.8765 -0.9726 -0.8922 -0.5438 -2.3792  0.6265 -0.5366  1.7281
#>  2.5567 -1.1786 -0.2495  0.2837 -0.7603 -0.4542 -2.3036 -0.4927 -0.9363  0.3087
#> -0.1621  0.5071 -1.0591 -1.7595  2.1782  0.2239  0.7759 -0.4475 -0.6190  0.3187
#>  1.0065 -0.0675 -0.4370  0.3801  1.9519  0.2597  0.0926 -0.5759 -0.1700  0.2583
#>  0.3327  2.0922 -0.4152 -0.9740 -1.9030  0.0113 -2.2927 -1.2652  0.5190  0.5542
#> -0.0418  1.3159  0.3025 -0.5086 -0.7685 -0.7787 -1.9117 -1.1234 -0.3496  1.1252
#>  1.6251 -0.8256 -0.9831 -0.9043  0.9272  0.1271  1.2555  1.7615 -0.0998 -0.0246
#> -1.5697  0.1196  0.0172  0.7749 -0.9459  0.2667  1.1276  0.8968  0.5840  0.2848
#> -0.1986 -0.1957 -0.0306  1.3735 -0.4565  0.1201 -0.6286 -0.0718  0.1805 -1.8618
#>  0.2692 -0.9121 -0.1380  0.2173  1.2318  0.7379  0.1031 -0.7796 -0.8223 -1.1750
#>  0.2278  1.5873 -0.1364  0.0327 -0.4581  0.8783 -2.4401 -1.0330 -0.5183  2.0320
#> -1.3198 -0.7424 -0.5712 -0.2734  0.2718  3.0456 -1.1786  0.0895  0.1181 -0.9443
#> -0.5322 -0.3504  0.2745 -0.5573  0.8543  0.3648 -0.8651  2.1433 -1.5275 -0.2002
#> -1.3068  0.3310 -0.7356  0.4657  0.5726 -0.9563 -0.3315  1.0597  1.6029  0.2015
#>  0.3874 -0.3826  1.1738 -0.0881  0.1640 -1.8429  0.0003  0.5842  1.2634  0.4697
#> -0.0248  0.6464 -0.2757  0.6083 -0.1099  1.9989  1.3440  1.1626 -0.3223  2.2685
#>  0.1416  0.9757  0.8712 -0.3587  1.0650  2.1883 -0.7116  0.6530  0.4568 -0.3229
#>  0.2284  0.9270 -0.2172 -1.6338 -0.5210 -0.5018  1.4707  1.7879 -0.2604  1.1650
#>  1.9640  1.4216 -0.0467  0.9610  1.1097  0.6898 -0.0147  0.0236 -0.2921 -2.0846
#> -0.2317  1.2851 -0.8367 -0.0200 -0.0839 -0.6280  1.1519 -0.1162  1.4542 -0.0099
#>  0.0485 -0.7216 -0.5248 -0.8564  0.9519 -0.4224 -0.0730 -0.4364 -1.3007 -0.9880
#> -0.9880 -0.6036  0.8302  1.2660 -0.0351 -0.2662 -1.0535 -1.1989  0.3729 -1.5479
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]