Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6845 -0.1712  1.3081  1.2684 -1.1102  1.3090  0.3491  0.2077 -0.2099  0.5273
#>  0.6877 -0.0285 -0.5750  1.1535  0.4940 -3.4792 -1.5337 -0.5021  0.3514  0.4223
#> -0.3797 -1.5574 -0.2571  0.3103  0.7432  0.4579 -0.0289  1.1621 -0.0409 -0.9711
#> -1.6734 -1.3484  1.9231 -0.9120  0.6925 -0.0373  0.5204 -0.8143 -0.9685  0.2159
#> -0.5833 -0.5180 -0.5682  0.1181 -0.1161  0.3408  0.3433  1.2732 -0.3653 -2.4554
#>  0.3729 -0.2463  0.9333 -0.3894  0.8755  2.0694  1.1587 -2.0185  0.1499  0.2296
#>  0.2525 -0.1656 -0.7373 -0.3479  0.8111  0.5471 -0.9376  0.8633 -0.4155 -1.1381
#>  1.4102 -0.5764 -1.3277  0.3061 -0.2797  0.1359 -0.4681 -1.2579 -0.3188 -0.1749
#> -1.2276 -1.6402 -0.7419 -0.1936  1.6359 -1.2223  0.5605  0.4904 -1.0268 -1.9830
#>  1.0592 -0.3976  0.0935 -0.6392  0.4701  0.3516  0.5422 -0.4805  0.0183  0.7572
#>  2.9965  1.9767  0.1496  0.0876 -0.6294 -1.2463  1.8568  1.8658 -0.6909  1.6425
#> -2.2318 -0.4833 -0.7528 -0.2415  0.1605  1.3775 -0.0002 -1.1932 -0.1005  0.9118
#>  0.5231 -0.0580  0.4567 -0.0026  0.4937  1.7036 -0.6040  0.1106  1.1755 -0.4016
#> -0.1684 -0.1202 -0.4288 -0.6561 -2.9231 -0.1937 -1.2120  0.2631  1.6012  0.0996
#> -1.5969  1.6336 -1.0470 -1.0563  0.3086  0.0691  0.1504  0.5752  0.3181  1.1277
#> -0.3189  2.7700 -1.0349  1.3245 -1.6107  0.3872  0.3364  0.5536 -0.5625 -0.9869
#> -0.3526 -0.2517  1.3068 -0.5095  0.4873 -0.6407  0.0945 -0.8266  1.5522  0.7554
#>  0.1872 -0.4783 -0.2262  1.4755 -1.1934  0.7311 -0.8412  1.1573  0.2011 -0.3140
#>  0.2675 -0.3426  0.3053  0.7920  1.5501 -0.0404 -1.5724  0.4514  0.7989 -1.9745
#>  1.0080  0.7554 -0.4102 -0.3956 -0.7089 -0.4950 -0.2527 -0.0547 -0.8147  1.0087
#> -0.7881 -1.6147 -0.5762  0.2254  0.4353 -1.1179  0.5278 -0.2844  0.5679 -1.7224
#>  0.8504  0.9866 -0.3843 -0.1701  1.0972 -1.1483  0.4951  0.2184 -0.5297 -0.6121
#> -0.5719 -1.3621 -1.3007  0.7111  0.8041 -1.9777 -0.4652 -0.0194 -1.1730 -1.3136
#> -0.8517  0.1981  0.5872 -0.3753 -2.3493 -0.5006  1.5745  0.0563  0.0966 -0.3500
#> -0.6816 -1.1440 -0.6109  1.0525  0.5972  2.1655  0.2301  0.8879  0.2362  0.3865
#>  1.8716  1.6667 -0.1955  0.7037  0.3882 -1.0549  0.9119 -1.3744  2.0275  1.1721
#>  0.4876  0.5830  1.4556  1.0690 -0.1026  0.9630  0.9382 -0.0257  0.5324  1.4023
#>  1.4150 -0.8537  0.6159 -0.2385  0.6404 -0.2706  0.8736 -0.8300  1.8060 -0.5219
#> -1.9454 -1.5039  1.5320  0.9147 -3.0444 -0.7615  0.7313  1.2363  0.4712  0.7645
#> -0.1074  1.2566 -0.3976 -0.2296  0.5327  1.1019 -0.9799  1.0185  0.0248 -0.1915
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]