Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2593  0.1159 -0.9020 -1.8032  1.0522  2.2095  0.0777  2.1607  0.1919 -0.3598
#> -0.2762 -0.6776 -1.9083  0.5292 -1.4635 -0.9966  0.5559 -0.8062 -2.3032  0.1476
#> -0.4144  0.3460 -0.2836 -0.0208  1.3240  1.2853 -1.3387 -0.3484 -0.1621 -0.0750
#>  2.9971  0.9586 -0.8404 -0.6645  0.5950 -0.4572 -0.1663  0.1383  0.0487 -1.6358
#>  0.3936 -0.4607 -0.3182  2.6824 -0.7219 -0.6918 -0.1331 -1.0874 -0.6864 -0.6072
#>  0.5401 -0.0273 -0.6819 -0.3168  0.6349  1.0389 -1.4477  0.1949  0.7804 -1.6476
#>  1.6121  0.8649  1.2054 -0.4848 -0.2868 -0.3280  0.2165  0.8385 -0.6341  0.8268
#> -1.2598 -3.3900 -1.2271  0.6309  1.5482 -0.1490 -0.5447  0.5157 -0.2908 -0.0922
#> -0.1845  0.5749  1.6493  0.6919  0.0522  0.6114  0.1768  0.4207  0.9160 -1.7018
#>  0.8184  0.2471 -1.8837 -0.7944 -0.5278  0.2788 -3.2493  0.1483 -0.1925  0.3156
#>  0.6691 -0.3010 -1.8523  0.6896 -0.5905 -0.5639 -0.0393  0.2319  0.9639  0.3071
#>  0.9340  0.9098  0.4182 -1.4189  0.5756  0.1952 -0.7651 -2.2794 -0.2255  1.3711
#>  0.3842 -1.9360  1.2339 -0.8450 -0.5630  0.1349  0.6199 -0.8794 -0.6633  0.3607
#> -0.2178  2.6740 -1.5309 -0.3295  1.3902 -0.3223  1.0909  0.0187  1.5924  0.6240
#> -1.5373 -1.4325  0.8728  0.8528  1.6297 -0.6426 -1.0480 -0.7568  0.6079  1.6763
#>  1.2175  1.2692  0.3117  0.6565  1.1348  0.2951 -2.0293 -0.4957 -0.1151  0.0584
#> -0.1832  0.4874 -0.1159  0.0726 -1.2951 -0.7855 -0.5981 -0.7825  0.7013  0.0753
#> -1.1525 -0.9410  0.2279 -0.6011 -1.2743  0.6037  0.0919 -1.1198  2.4250 -0.1167
#>  0.2174  1.2286  0.3348 -1.5375  0.0590 -1.0261 -0.3673 -0.5179  0.4507 -0.3146
#> -0.0564 -0.3805  0.8084  0.2695  1.2202  1.8553 -1.3964 -0.5084  0.2625 -0.0176
#> -0.9648  0.0569  0.8018 -0.7774 -0.0763  0.9888  0.5374 -0.4443  0.2068 -1.3280
#>  1.2186 -0.3835 -0.1851  1.4458 -0.5896 -0.2238  0.5283 -0.3284  1.5946 -0.5501
#> -0.7188  0.0921 -0.4408  1.0498  1.6404  0.1325  0.6790 -0.3601 -0.1098 -1.0207
#> -2.0089  0.4187 -0.7454  1.2069 -1.4442  0.9462 -0.8082 -0.5516  0.5876  0.1151
#> -0.9291 -1.2564 -0.9788  0.9229  1.8143  1.0517 -0.7989  0.2878  0.1303  0.6686
#> -1.6103  0.2194 -0.3838  1.6545  0.0120 -2.1509  0.3959  0.6067  0.8733  0.1450
#> -0.7525 -1.1888  0.3276  0.3124  0.2768 -0.8637 -0.6538 -0.4939 -0.3123 -1.1353
#> -0.6832 -0.6214  0.6294 -0.2167 -1.7742  1.0076  0.5958 -0.3364 -0.1314  0.0455
#>  0.3465 -1.1115  0.0365 -0.1782 -0.7152 -0.6917  0.9569  0.4203 -0.3816 -1.0776
#>  0.1527 -0.8623  0.0388  1.2774  1.4252 -0.8781 -0.8300  0.3672 -0.6411  0.9807
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]