Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.2101 -0.0565 -1.3677 -0.4233  0.3615  0.9731 -0.1037 -0.1580 -0.3397  0.5640
#> -1.1427  1.3673  0.1043 -0.2670  0.5033  0.7547 -0.6037 -0.2086  0.9792 -1.9666
#> -0.0291 -0.6105 -0.1438 -0.4195 -0.5629 -2.0582 -0.8392  0.0430  1.0054  0.1465
#>  1.5558  0.6952  1.9419  0.8911  0.5152 -0.4564  0.9672 -1.3418  0.5833  0.6140
#>  0.3049  1.0031 -0.2358  0.4261  0.5660  0.0820  1.4724  1.4078  0.3538  2.7373
#>  1.2470 -0.3256  0.6976 -0.3804 -0.9774 -1.1255 -0.7667  1.4196  0.8691  0.6529
#>  2.0538 -1.1625 -0.1502  0.9063 -1.2119  1.9162  0.7885  1.2122  2.6925 -0.2571
#>  1.7373 -0.9901  0.4735  1.4269  1.7583 -0.5239 -2.8715 -1.8356  2.4838 -0.3631
#>  0.4044 -0.9512 -0.6617  0.7451 -0.3015  0.5750  1.9216  0.1079 -1.8550  0.0614
#> -0.1259  0.2847  0.0976  0.2693  0.0267  0.2133 -0.3876  0.3737 -1.0479 -0.9175
#>  1.2937 -0.1813 -0.4156  0.6838  0.6187  0.1365 -0.1408  0.2472  1.0293 -0.3818
#> -0.6774  1.6822 -0.2718  0.5077 -0.4633 -0.4896  0.3929  0.4224  0.6252 -0.4212
#>  0.1949 -0.3620 -0.7581 -0.0879  0.7704 -0.4587  0.4659 -0.9957  0.6070 -0.3647
#>  1.0914  0.9573 -2.2380 -0.2134  0.8831  0.5413 -0.3525 -2.5034  1.8112  0.2804
#>  3.3989  1.5356 -1.3210  0.5429 -0.8578  0.0951 -0.3932  1.1777  0.3696 -0.7617
#>  0.4994 -1.0923 -0.3278  0.7958 -0.3548  1.4904  0.0329  0.8167  2.6736 -0.0384
#> -0.4962 -1.8337  0.9297 -0.4282 -0.2736  1.2315  1.5122  0.7143  1.2302 -0.0844
#> -1.2232 -0.1915  3.3269  0.2754 -0.1095  0.6258  0.9706  1.9142  2.4648 -1.4424
#> -0.5209  1.4843  0.6086  0.4077 -0.5548 -1.0616  0.0744  0.7892  0.9637  0.2813
#> -0.0331  0.0741  0.6329  0.0045 -0.0467 -0.6298  1.5242  0.7667 -0.0994 -0.1061
#>  1.0812 -0.3659 -2.2411  0.9444 -0.9279  0.3543 -0.3306 -0.8202 -0.7635 -1.1227
#> -0.7266  0.8689 -0.5348  0.4508  1.1897  0.0749 -0.4812 -0.0461 -0.6895  0.1177
#> -0.0150  0.1879 -1.7809  1.2410 -0.6120  0.8642 -0.9951  1.5303 -0.4821  0.7366
#> -1.3853 -1.9050 -0.5751  0.0946 -0.6020  0.8358  1.0876  1.4303  1.1693  1.2349
#> -0.1874  0.3519  0.0995  0.5475  0.0005 -0.5307 -0.6020 -0.2939  0.1397 -0.8862
#> -2.2333  1.8852  0.8024 -0.1031  1.4266  0.9496  0.9075 -0.6254  0.8149 -0.0435
#>  0.7610 -0.3101  2.2953  0.1249  0.0791  0.5886 -1.1412 -1.0214 -0.2298  0.0040
#>  0.0775 -0.8377 -0.7380 -1.0898 -0.4687 -1.5227  1.6770 -1.5615 -0.5526  0.1900
#>  0.0766 -1.4466  1.3278  0.3565 -0.4945 -0.5097 -0.6081  0.7128  0.2000  0.3932
#> -0.1197  0.2827 -0.0333 -0.1160 -1.6641  0.5036 -1.0999 -0.7895  0.7529 -0.1374
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]