Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.3586  0.6965 -2.6540  0.6609  0.5237  0.7659  0.1529 -1.6797 -1.3100  1.2673
#> -0.8376 -0.0556  0.4773  0.2032  0.8116  1.4610  1.2885 -1.2096 -0.9051 -1.1534
#>  0.8358 -0.7316  1.2537  0.2808  1.4217 -1.0890  0.2329 -1.1222  0.9803  0.9355
#>  0.8973 -0.1138 -0.5551  1.0494 -0.3595  0.3745  0.8613 -0.4677 -0.1475  0.3478
#>  0.4602  0.1800  1.1627 -0.5909 -2.1186  1.2000 -0.2269 -0.3497 -2.3326  1.8628
#>  0.6558 -0.7621  0.5142  1.9103 -0.5231  0.1388  1.5244 -0.6083 -0.6603 -0.3601
#> -2.0535 -1.5253  0.0748  0.8082  0.6737 -0.2682  0.5920  0.3276 -0.5398 -0.5522
#> -0.6779 -0.6016  0.7459 -0.1228  1.0621 -0.7890 -0.2115 -0.8682  0.8162 -0.0908
#>  0.3249 -1.8841  0.4463 -0.0036 -0.0447 -1.2093  1.1484  0.9129 -0.7548  0.0418
#> -1.6198 -0.8874  0.0864 -0.8304  0.0022 -0.3713  0.1463 -0.3616  1.6684  0.7120
#> -0.1105 -0.7962 -1.7214  1.4298  1.1213  0.7400 -0.5118  0.5115 -0.1347 -1.6713
#> -1.8657  0.8858 -0.0606 -2.1592  0.6384  1.1856  0.2623 -0.3871 -0.5221 -0.2658
#> -0.8563  0.1255  0.3494 -0.3440 -0.4772 -0.4021 -0.0340 -1.3362  2.0817 -0.1922
#>  1.6905 -0.3768  0.4615  0.3909  1.0997  1.0534  0.2983  0.3097 -1.0205  0.6387
#>  0.8752 -0.2008  1.4267  1.2017  1.1010  1.3071  0.1721 -0.1860  2.1848  0.3076
#>  0.3970  0.0048  1.0685 -0.1124 -0.5819  0.4622  0.5357 -0.8204 -0.2079  0.4216
#>  0.2670 -1.6869  1.2360 -0.3588 -0.2619 -1.0495 -0.8163 -1.4769 -0.6664 -0.1094
#>  0.8021  0.8807  1.6722 -1.8122 -1.6679  1.1314 -0.0213 -0.0956 -1.2666  1.5095
#> -0.1718 -0.0099  0.9165  0.1241 -0.4906 -0.3091 -0.8356  0.7761 -0.4072  1.1718
#>  0.2084 -0.5740 -0.5879  0.3436 -1.5396  1.3531 -0.1746 -1.4856 -0.6851  0.5867
#> -1.1500  1.8780 -0.5323  0.2530  0.6276 -1.3056  1.7263  0.2283 -0.6081 -1.1714
#> -0.8343 -1.0709  1.0093 -1.0831  1.0709 -1.5884 -0.4683  0.7319 -0.6277  0.9844
#> -1.4647 -0.8087 -0.0709  0.5082 -0.4998 -0.5567  0.3815 -0.5008 -1.5880  0.0074
#> -0.8288 -1.8457 -0.7708  0.5185 -1.2375 -1.4997 -1.0605  0.5237 -1.6350 -0.1572
#> -0.5983 -1.5672 -2.4124 -0.9585 -0.5921  0.2895 -2.6157  0.1988  1.0004  0.1302
#> -1.0348  0.5678 -0.8153  0.6191 -0.4467  1.3654 -0.1617  1.7169  0.6586 -0.6412
#>  0.6520  1.4645 -0.7130 -0.9065  0.1061  0.4651  0.9105 -0.4109 -0.3554 -1.1730
#> -0.2883  0.0279 -1.5919 -0.5610  0.0405  0.1629  0.1320  0.5951  0.5377 -0.4243
#> -0.7573 -0.6563  0.6631  0.5499  0.7679  0.7358 -0.6224  1.1282 -0.8330 -2.3278
#> -1.0350 -0.2383  0.1435  1.2866  0.5686  0.6075 -0.4656  0.1706 -0.1539 -2.0385
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]