Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0872 -0.8618 -0.4917 -0.3557 -0.2295 -0.4905  0.3805  0.1362  1.3677 -0.3167
#>  1.2762 -0.4942  0.3155 -0.2721 -2.3886  0.0871 -0.4383  0.4394 -0.2033  1.8296
#> -0.3688  0.3517 -0.0161 -0.9117 -1.0600  0.2377  0.6998 -0.6588  1.5658 -0.3717
#> -0.5524  0.3657 -0.2507  0.3078  0.3617  1.0374 -0.2388 -0.1693 -0.5214  2.0421
#> -1.1887 -1.9349  0.4089 -0.3539 -0.0699  0.1567 -1.1966 -1.1282  0.9284  1.3242
#>  1.7873  0.4877 -1.9776  0.2955  0.1627  1.0587 -0.7575  0.7975  0.0875 -1.8774
#>  1.2415  0.2069  0.7864 -0.3458 -0.9302 -0.2492  0.4602 -0.6476  0.3990 -2.2739
#> -0.8379  1.9926 -0.6551  0.1276  2.2886  0.0243 -0.5842 -1.0394  1.1570 -0.4860
#> -0.6681  0.3028 -0.2203 -3.1702  1.8617 -0.3539  0.6323  1.2468 -0.4560  0.4450
#> -0.1548  0.1215 -0.4417 -0.3612  0.3744  2.0553 -0.1899 -0.1313 -2.0105 -0.9968
#> -1.0483 -0.4554  0.2843  0.6560 -0.6843 -0.8401  0.2127 -1.3645 -1.7163  0.2920
#> -0.0865  1.9200  0.7283 -0.7327  0.3933  0.4425 -1.3140  1.6735  0.1454  0.3391
#> -0.4222  1.1870  0.3275 -0.7085 -2.2324 -0.2861 -0.2566  0.1473 -0.4390  0.0027
#>  0.4007  0.6735 -1.8802 -1.7103  1.0108 -0.6069  0.0339 -1.3871 -1.6581  1.4313
#> -0.3502  0.8666  0.9179  0.8195  0.7957 -0.9303 -1.2433  0.8688 -1.5553 -0.9827
#> -0.8594 -1.2345 -0.2164 -0.7078  0.3839  0.9325 -0.7251  0.2077 -2.6054  0.9785
#> -0.0167 -0.9590 -0.3750  0.3500  0.7143  0.1473  0.0110  0.5855 -0.2780 -0.9232
#> -1.6655 -0.8909 -1.0215  0.8069  1.2847 -0.3267 -1.5253 -0.4504  1.9550  0.9390
#> -0.2259 -0.8533  1.7425  0.0018 -0.3497 -0.6684  0.8148  0.5712 -0.7346 -0.1812
#>  0.1656 -0.3429 -0.1010 -0.3059 -0.5708 -0.6328 -0.1202  1.5572  0.0126  0.4234
#>  0.1837  0.5654  0.4053 -0.0561  0.2566  1.1442  1.1069  0.0794 -0.3718 -0.8426
#>  0.0232 -0.0923  0.3881 -0.2783 -0.5807 -0.0653 -0.0274  0.0800  0.0722 -0.0220
#> -0.9734 -1.9133 -1.7707 -0.2906  0.2293 -2.1018 -0.1082 -0.2562  1.3760  0.7147
#>  0.0803  1.1447 -0.7337  1.1199  1.4344  1.3985 -1.4204 -0.0673 -2.0310 -1.1505
#>  1.3805 -0.6649 -0.0156  0.1017  1.0172  0.6288  2.1409 -0.6724 -1.7355 -0.7629
#> -0.4547  0.7045  2.1381  0.5145  0.9264 -0.9438  0.7913  0.5474  1.0169  0.1236
#>  0.0232  1.5235 -0.5069 -0.9775 -0.4740 -0.7550 -0.9927 -0.1468  1.3025  0.2862
#>  0.0488  0.0844 -2.2620 -0.4000 -1.0713 -0.9045  0.5262  1.4833 -0.4379 -0.0147
#> -0.0303 -0.1325  1.4369  0.8802 -0.0811 -0.4382 -0.7839  0.4961  0.2705 -0.4928
#>  0.6936  0.5515 -1.5059  1.9559  0.0912 -3.2338  0.8297 -1.2601  0.6839  0.2927
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]