Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.1366  0.6110  0.2410 -0.3997  0.3368  1.4023  0.3254  0.9833 -0.2843 -0.6363
#>  1.9000 -0.8032  0.2642  1.4124 -0.4552 -0.5750 -0.6366  0.3477  0.8389  0.4696
#> -0.0724 -0.0307 -0.0800  2.0675  2.2592 -1.4420 -1.5985  0.7073  0.4545 -0.4313
#> -0.9356  0.4262  0.4126  0.7049  1.3422 -0.9859 -0.6891  0.5848 -1.7164  0.4585
#> -0.3673 -0.1273  0.6399  1.1590 -0.7310  0.7701  2.0777 -0.4629  0.3559  0.1687
#>  2.0092  0.0020 -0.8157 -0.1171  0.7998  0.3678  0.3187  1.2725 -0.6646 -1.0566
#> -0.6043 -3.3359 -1.2545  0.6867  0.9082  0.4189 -0.5015 -0.5433 -0.9813  0.2970
#> -0.9772 -0.5165  1.5061  0.5681  0.5637 -0.9267  0.3452  1.4643  1.0392  0.0881
#>  0.2838 -0.0571 -0.9088  1.1070  0.7893  1.1868  0.6994 -0.3255  0.0619  0.3550
#> -1.7271 -0.0256 -0.3673 -1.3645 -0.2549 -0.7054 -1.6153 -1.5790 -1.5670 -1.0805
#>  0.5105 -0.3926  0.0426  1.9156  0.7488  0.0227 -0.8265 -1.3056  1.5659  0.1760
#>  0.7629  0.2862  0.3633  0.3368  0.7987  0.9218 -0.3636  0.0946  0.2705 -0.3506
#>  0.3004  1.3932 -0.0551  1.6017 -1.0191  0.7104 -1.9234  0.1651 -0.7042 -1.5648
#> -0.0815 -1.5769  1.7706  0.3179 -0.6932  0.3487  2.0348 -0.1645 -0.6224 -0.6778
#> -0.9684 -0.3468  2.2937  0.9626 -0.0906  0.8138  0.3446 -0.1205  0.6277  0.3424
#> -1.6580  1.3873 -0.7051 -2.2273  0.8700 -1.0555 -0.0122  0.4048  1.3241 -0.3706
#>  0.2948 -0.6911  0.2442 -0.8249 -0.2315 -1.1166  0.3046 -1.4878 -0.2432 -1.1877
#> -0.3643  0.6875  0.1614  0.1397 -1.3068 -2.6478  0.0072  0.1771  0.1312 -0.9675
#>  0.7427  0.1790 -2.6175  0.2624 -0.4213  0.0606  0.7566  1.3674  0.2692 -0.8529
#> -0.9472 -0.8870 -0.6982  0.8684 -0.1012  1.4703  0.5410  0.5537 -0.4370 -0.6598
#>  1.7424  0.8794  1.1930  0.2636 -1.7628 -1.1082 -0.1143 -1.0479 -0.0518 -1.4218
#>  1.5426  0.8992  1.0214  0.4376  0.2994  1.0880 -2.1317 -2.1316 -0.4248  0.4985
#> -0.3678  0.9538  0.1144 -1.6395  0.0444  0.4442  0.5311 -3.1868  0.0244 -1.0801
#>  1.5811 -0.5764 -0.5072 -0.2563 -0.9942  0.3416 -1.5208  1.0410  0.8878  1.2592
#>  0.4788 -0.6750  0.6475  2.0863  0.5400  0.2120  1.2979 -1.4474 -1.6494  1.3109
#>  0.4618 -0.0769  1.4044 -0.1041 -0.9830 -0.4988 -0.8578  0.4417  1.4737 -0.8339
#>  0.5027  0.6246 -0.1439  0.2599 -0.4849  1.2455 -1.1493 -0.8775 -0.3299  0.1329
#>  1.3674 -1.7401  0.1440  1.2061  0.0473  2.1162  1.2439  0.8061 -1.3138  0.8389
#> -1.0217 -0.1705  0.6582 -0.7615 -1.3164 -0.3418 -0.6533 -0.0808 -0.3087 -0.9574
#>  1.1700  0.1396 -2.2015  0.0906 -0.1953 -1.0502  0.5425  0.4155 -0.0254 -1.4071
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]