Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.4215 -0.1248  0.0978  0.4073  0.5933  1.4496 -0.0964 -0.8525 -2.7767 -0.3493
#>  0.2647  1.0979  0.2398  0.2326 -0.2153 -0.3083  0.0235  0.2210  1.4170 -0.6143
#> -0.6883 -1.0816 -1.1880 -1.0584  0.2530 -0.4198 -0.1421 -1.5152  3.3705  0.3069
#> -0.6046 -0.5076  0.6295 -0.3812  0.2029 -0.1448 -0.7608  0.5358 -1.0906  0.6757
#> -0.6726 -0.5005  0.3836 -0.0229 -0.5334  0.0937  0.2925 -0.1046  0.5940  0.2723
#> -1.8537  0.2738 -1.5235 -0.3899 -0.8633  0.4229  0.3456 -0.9742  0.5937  0.1627
#> -1.2644  0.8549 -0.1310  0.7896 -1.6484  0.3086  1.6735 -1.7928 -1.6254 -0.5077
#>  0.9677 -1.5027 -1.3233 -1.3429  0.6244  0.8478 -0.6308 -0.4818 -0.0319  0.1908
#>  1.9923  0.8419 -0.7766  0.2319 -0.2262  1.2328  0.7727  1.4165  0.9517 -0.1866
#> -0.9571 -1.2268 -0.0484  0.2679 -0.7240  0.1252 -0.2563 -0.0927 -0.8921 -0.4247
#>  0.8387 -0.1107  0.0541 -1.1886 -1.5733  0.0258 -0.6941 -0.3571 -0.4026 -0.9012
#> -0.6957  1.9534  1.1227 -1.5728  1.5001 -0.7314  0.7597 -0.0883 -1.0511  0.8863
#>  1.2533 -0.2862 -1.9500 -0.3510 -1.6253 -0.7012  0.8809 -0.6299 -0.2286 -0.4196
#> -1.7844 -1.4450  1.2321  1.6770 -0.8599  1.1716  0.1963  1.0516  1.3614  0.8189
#> -0.9473  1.2796 -1.4951  0.3231 -0.6088  0.6025  0.1390  0.8439 -0.1454  0.1448
#>  0.9653 -0.5566  0.2071  1.0706  0.4466  0.7274 -1.6157 -0.8822 -0.8436 -0.0878
#>  1.0236  0.2441 -1.4437  0.6644  0.4349 -0.1164  0.2076 -0.8515 -0.4572 -0.2687
#>  0.6560  0.4421 -0.3337  0.4334  0.1480 -0.1450 -0.8070 -0.0785  0.3452  1.0769
#> -0.4819 -0.8506  0.8330  1.6701 -0.6816 -0.6266 -0.1618 -0.6863 -1.6081 -0.1473
#> -2.2439  0.5751 -2.0735  0.0864 -0.1094 -0.5814 -0.0469 -0.9635 -0.9768  0.4265
#>  0.5012 -0.5237  1.5663 -0.0271  0.1237  0.2310 -0.3217 -1.0692  0.1254  0.5454
#>  0.4278 -1.5559  2.3693 -1.2699 -1.5285 -1.9769 -0.9743 -0.5747 -1.9050 -0.4878
#>  0.2521  0.1377  1.9269 -0.5543  0.1600 -1.9292 -0.5231 -1.6363  0.1030  1.9756
#> -0.6567 -1.0688  0.8542  0.4630  0.3605 -0.8112 -0.7480  1.4956 -1.1376  0.8224
#>  0.8414  1.2144  1.4630 -2.4498 -0.6859  0.8373 -0.8484 -0.5445  0.1342  1.0418
#> -1.3823 -0.1824 -0.1788 -0.0078  1.3032  0.8950 -1.9608  2.0400 -0.1009  0.7249
#>  0.1904  0.6298 -0.4721 -1.2099  1.0578 -0.0675  0.5476 -0.4035  1.1552 -1.1926
#>  0.4956  0.1584  1.8230  1.1876 -1.9310  1.9088 -0.0330 -2.9755 -1.6179  1.3649
#>  0.6434  1.2146 -1.3820 -0.2177  0.5862  1.5628  0.1570  0.5096  0.8475 -1.2030
#>  0.7800 -0.0862  0.2833 -0.5182  0.6426 -0.5894  0.1265  1.3971 -0.6610  1.3351
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]