Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6260 -1.0995 -0.4971  0.1750 -1.8210 -0.1415  1.3171 -0.1994  0.6572  1.9216
#> -0.6927  1.3815  0.1094  0.6785  1.1091  0.8024 -1.3445  1.0831 -0.7736  0.5036
#>  0.1567 -0.8713  1.5714 -0.5407  0.1122  0.4783  1.0268  1.0084 -0.8163  1.2780
#>  1.4154 -0.4261  0.7828 -0.5395 -1.1368 -0.3682  1.2258 -0.3403  1.0744  0.0958
#> -0.4195  0.3252 -1.5196 -0.8701 -0.6063  1.4380  1.0236 -1.1626  0.6212 -1.2566
#> -0.6176  0.3175 -1.4109  0.2480 -0.2166  1.1713 -0.4994 -0.9083 -0.2509  1.5100
#>  1.7123 -1.1399 -1.3216  0.1461  0.8217  1.1355  1.1459  1.0073 -0.0751 -0.0507
#> -0.4299 -0.8967  0.8038  0.4031  0.7690 -0.3689 -0.5200 -0.7683  0.2461  0.7234
#> -0.6931  1.0451  0.4359  1.1811 -1.8780 -2.3409 -1.7138 -0.2913  1.0316  1.3277
#>  0.2119 -0.5071 -1.3362 -0.3050  1.0461 -1.4958 -0.9155 -0.9414 -0.0435  0.6368
#> -2.1589  0.1163 -0.4935  0.2286  0.9554 -1.3972  1.5444  0.1991  0.0638  0.6397
#> -1.1641  0.5619 -1.5000 -0.3791  0.9717 -0.8191 -0.0221  0.7757  0.7219 -1.4067
#> -0.1606 -0.4569 -0.6061 -1.2399 -0.3418  0.7780  0.6864 -0.0868 -0.6226 -0.2388
#> -1.9237  0.1619 -1.0250  1.2267 -0.4448  1.2083  0.9941  0.9578 -1.3263  2.0270
#> -0.6194 -1.9218 -0.4992  1.3092  0.9719  0.3517 -0.1008 -0.6197  0.4787 -0.1920
#>  1.7280  0.5517  2.0961 -1.7643  1.5779  1.0179 -2.3802 -0.6286 -1.1954 -0.5106
#>  0.7553 -0.3108  0.4194 -2.0602  0.2312 -0.0098  0.3670  0.3171 -0.7668 -0.6816
#>  0.0525 -0.1413 -0.7402 -0.4597  0.0575 -0.1193 -0.2358  1.6632  2.3393 -0.7701
#>  1.4772 -0.1118 -0.8966  0.4734 -1.2504 -0.5450  1.2414  1.8464  1.1482  0.2023
#>  0.1703  0.8522 -1.0388 -0.3266  1.2277  0.7425  0.1207  1.9689 -0.4613  0.0137
#>  0.5355  0.0988 -0.3069  0.1915 -0.2388 -0.2445  0.1588 -0.9411  1.4155 -1.0429
#> -0.2891 -0.8933  1.4494  1.2528  0.2629  0.5084 -0.8058  0.1124  0.2035 -0.1694
#> -0.3916 -1.3999  0.2822 -0.5411 -0.8511  0.6773  1.4824 -0.6893  0.8251 -0.0718
#>  0.0780 -0.4198  1.8894 -0.3235  0.2212 -0.4860 -3.2155  1.9779 -0.0954  0.6975
#>  0.7827 -0.1494 -0.2020 -0.0622  0.4653 -0.9267 -1.2845 -0.0330 -1.7499  1.6676
#>  1.2600  0.5939 -1.8382  0.9103  1.5897 -0.0747 -1.6384 -0.4598 -0.8600 -0.1099
#> -1.3482  0.1594  1.5913 -0.1679  0.6632  1.3802  0.2372 -0.0308  1.4491 -0.1196
#> -0.2874 -0.3731  0.8968  0.2808  0.3060 -0.2472 -0.1775 -1.2434  2.6520  0.3564
#> -0.3168  0.0523  2.6033  2.4228  0.8169  0.9628  0.1042 -0.4807 -0.2805  0.4566
#>  0.5565  2.0010  1.3718 -0.3484  2.0767 -0.3268  1.9338 -0.9392 -0.2639  1.2489
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]