Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.1508 -0.9148 -1.4818  0.7603  0.1084 -1.3326 -1.8967  2.0008 -3.0378  0.1165
#> -0.5449  1.4285 -0.1145 -0.5393 -0.6967 -0.0812  1.9611  0.4564  0.1551 -0.1245
#> -0.8110 -0.0716 -0.3879  0.8564  1.0265  0.3362 -1.3623 -0.4954 -0.3600 -0.0097
#>  0.5763  1.3263 -1.6126  1.2671  0.2122 -1.0259  1.3432 -1.6347 -0.0147 -0.1074
#> -0.4984 -1.0555  0.3819 -0.3955  0.2523 -0.5617 -0.0394  0.5232  1.2088 -0.3484
#>  0.9383 -0.7045 -0.0316  0.3806  0.0499 -1.5126  0.5729  0.1867  0.0579  0.8351
#>  0.1830 -0.7744  1.6942  0.9461  0.0728 -0.0614  0.5872  0.0632 -0.9309 -2.1191
#>  0.8738  0.0577 -0.4845 -0.6069 -0.9145  0.4113 -0.7552  0.0782  2.0549 -0.0434
#> -1.1142 -1.2634 -0.0691  0.6134 -1.0625 -0.6503  0.9659 -0.4488  1.4190 -0.0953
#>  1.0406 -0.0695  0.5490 -0.1311  1.5533 -1.0149  2.2714 -0.3975  1.1947 -1.2175
#>  1.6299 -0.1926  1.5673 -0.0108 -0.2843 -0.9989 -1.2934  0.8592 -1.0157  0.4992
#> -1.3480 -0.0183 -0.6289  1.6028 -0.9109 -0.0597  1.2935 -0.9747  2.0829 -0.2132
#> -1.1770  2.0901 -1.4201 -0.9010 -0.3314  1.2153  1.3351  0.0983  0.4067  0.5606
#> -1.0161 -1.0768 -1.3813  2.1842  0.8168  1.0786 -0.1853  0.0382 -1.6256 -0.8171
#> -0.0428  1.2260 -0.5889  0.0591  0.9376 -0.5773 -1.2854 -0.8875  0.5051 -0.3615
#> -0.4072  1.4304  0.3115  0.1562 -0.0022 -0.7926  0.0537  0.2938  1.7493  1.6578
#>  0.0156  0.7846  2.0366  0.0792 -0.7649  0.2083 -0.1853  1.5883 -0.7972 -1.0433
#> -0.5006  0.2574 -1.6691 -0.2240 -0.0055 -0.8321  0.0941  1.1591 -0.8200  0.5546
#> -0.8754 -0.2709  0.2126 -0.3967 -2.0004  0.8734  3.0311  0.0453  1.0128  0.4603
#> -1.3171  0.7839 -0.4041  0.4352 -0.0243  1.0074 -0.9826 -1.2591 -1.4942 -1.2619
#>  0.5486 -1.4994  0.0882  0.1817 -0.6501 -0.7570  1.4097 -0.1503  2.1973 -0.0230
#> -0.6138 -0.4088  1.0329  0.7778 -0.8485  2.1148  0.0598  1.4656 -1.8008 -1.3633
#>  0.0140  0.0309 -0.7760 -2.3346  1.4836 -0.0042 -1.4592 -0.6883 -1.2901 -0.3907
#>  0.2908 -1.6161  1.2939 -0.4371 -0.8323 -2.0656  0.9425 -0.3133  0.5116  0.8820
#> -0.2538  1.5763 -2.2783  1.8848  1.3728 -0.1537 -0.0354  0.6886 -0.0733  1.5667
#>  1.9022 -0.0003  0.1805  0.0697  0.3670 -2.1254  0.2216  0.8567 -0.3255 -0.8427
#>  0.0747 -0.4922  1.0757  2.3951  1.5412 -1.9354 -2.5913 -0.3408  0.1932  2.1089
#> -0.3980 -0.8168  1.1755  0.3814  0.7628  0.9047 -0.9662 -0.9064  0.0434 -0.0345
#> -2.3642  0.8151 -0.7125 -0.1834 -0.0947 -0.1931 -1.0148  1.5082 -0.7152  0.5739
#>  2.2694 -0.6544 -0.0828  0.8151  0.4529 -1.5603  0.5617  0.6106  0.4884 -0.5244
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]