Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7187 -0.3244  0.7659  0.8373 -0.8845  0.6495 -0.6381 -0.4606  1.2435 -1.6978
#>  0.9979 -0.2491  0.2345 -1.3009  1.9568 -0.4562  0.6376  0.4915 -1.2215 -1.6580
#> -0.6844 -0.5426  0.3084  1.5760 -0.6588 -2.1630  0.2627 -0.4160 -0.5117  0.2135
#>  0.3169  0.3609  1.2833  0.3036  0.6665  1.0967 -0.0303  0.7486  0.5686  0.4158
#>  0.2306  1.8967  0.8259  1.5729  1.8104  0.6555 -1.5913  0.6996 -0.1654 -0.5114
#> -1.0692  1.1055 -0.5377  1.4071 -0.8647  0.1244 -1.0282  0.2568 -1.5680 -0.1393
#> -0.6046  0.7667 -1.6772  1.1606  0.4777 -2.3529 -0.1501 -1.4094 -1.7983 -0.9051
#> -0.7619 -0.0029 -0.1796 -0.2940  1.8388 -0.8380  0.2018  0.7016 -0.8688 -2.4009
#> -1.8086  1.8967 -1.5637 -0.2204  1.6751 -1.0015  1.1693  0.7906 -0.3822  1.7329
#>  1.4725 -0.6061 -0.2050 -0.0768  1.7819  0.3412 -0.6747 -0.3693  0.0867  0.4740
#> -0.6412  2.2922  1.2281 -1.7215 -1.3193  2.2923  0.1402 -1.6516 -0.0381  0.0863
#> -0.2669 -0.1013 -0.6129 -0.3917 -0.2792  1.2133  1.2947 -0.4704  0.7595 -0.2744
#>  1.1736 -0.3615 -0.2629  0.2379 -0.5865  0.1414 -0.2795 -0.8123 -1.7597 -0.6262
#>  0.4954  0.1204 -1.6544 -1.1294 -1.2929 -2.0578  0.5065 -0.5858  1.5998 -0.8123
#>  0.2959 -0.1258 -2.6491  0.0631 -0.1668 -1.0981  2.2418  0.2989  0.3636 -0.5066
#>  0.7379 -0.9592  0.8988 -1.0908  1.5554 -0.2077  1.0647 -0.7770  0.4530  1.1751
#> -0.2778 -1.1532  1.2760 -3.5040 -0.0005 -0.3603  0.9134  2.0611 -0.8036  0.6425
#> -0.6740  0.1577  1.8061  1.1061  0.9329 -0.5590  2.3005  0.0310  1.2542  0.0895
#> -1.0137  1.0049  0.4124 -0.6285 -0.6245  0.1720 -1.4034 -0.0872  0.5298 -0.8770
#> -1.0893  0.5695  3.0346  0.1162 -1.1459  1.2540 -0.5610  0.7002  0.1573 -0.9867
#> -1.8025 -1.5458 -1.5891 -0.4316  0.0977 -0.6104  1.6966  0.7457 -0.6240 -0.5262
#>  1.7906  2.1319  0.1244  0.1875  1.3031 -1.1766  1.2820  0.3857 -0.8083  0.3045
#> -0.0009  1.7638 -0.1353 -0.1419  0.7145  0.3556 -0.9936  0.7693 -0.1389 -0.4627
#> -0.3850  0.5996 -0.6193 -1.3102 -0.3721  1.7043  0.6791 -0.3784  0.1186 -1.6015
#> -1.3421 -1.0151  1.8876  1.2676 -0.8958  0.0190  1.1965  1.3466 -0.3828  0.5281
#>  0.5278 -0.1308 -0.4984 -0.3189  1.1309  1.1431 -1.1896  0.0422  1.2985 -1.4102
#>  0.0347 -0.5245  0.5017  0.9683  0.2702  1.1783 -0.1834 -0.2800  0.2806  0.4030
#>  0.0023 -0.6356 -0.1224 -0.5354 -1.0147 -1.2788  0.8715  0.0469 -0.6917 -0.7731
#> -0.8988  0.6037  3.0269 -1.7121 -0.6898 -1.0551  0.5365 -0.2845 -0.9564 -1.6434
#> -1.1315 -1.2622  0.7119  2.0354 -0.8561 -0.9641 -0.6925  0.4130  0.1010 -1.7763
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]