Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.9655  0.0280  3.0640 -1.1096  1.1818  0.3091  1.6447  0.0437 -0.7412 -0.7767
#> -0.1597 -1.1650  1.4714  0.2154  0.5509 -0.5013 -0.2756  1.7383 -0.9578  0.7899
#>  0.8592 -2.0213  0.3832  0.0033  0.9076  1.1177  0.1962 -0.2574 -0.2863  0.0073
#>  1.6825  0.6605  1.6396  0.2569 -1.0272  2.8305  0.7840 -0.1924 -2.4232  0.5329
#> -0.0706 -0.8433 -1.0007 -0.4496 -2.2676 -2.2697 -2.1869 -0.6914  0.0557 -0.5161
#> -0.8105  0.9743 -1.0721 -2.2147 -0.4835  0.0968 -1.3401  0.8364  0.5004  0.7030
#>  0.3944  0.2293 -0.3300 -1.8222 -2.1099  0.3685  0.4804 -0.1108  1.2988  1.4071
#> -0.8889  2.1099  2.0493  0.3329 -0.2776 -0.6338 -2.0463  1.3837  0.8915 -0.1149
#> -1.2623  0.1239  0.7921 -0.1659  0.1906 -0.2011 -0.8083  1.3162  0.1545 -1.1293
#> -0.8824  1.2584 -0.0693 -0.0575 -0.6922  1.1827  0.1272 -0.8937 -0.8605 -1.1068
#>  0.0788  0.2548  0.2808  1.4719 -0.4855  0.8606  1.6259 -1.4500  0.1585  0.8331
#> -1.2274  0.7149  0.6245 -0.6074 -1.3048 -0.0259  1.6413  1.7299 -0.0083 -0.9926
#>  0.1434 -0.4372  1.1728  0.8295 -0.5320  1.4726  0.6554 -1.8636  0.3479 -1.0321
#> -0.5040 -1.8321  0.6672  0.3625 -0.3564 -0.7908 -1.1801  0.2534 -0.8224  0.1730
#> -0.3610 -0.8653  0.4118  1.3212 -0.0333  0.2321 -0.7578 -0.0787  1.2303  1.6736
#> -0.6690 -0.5533  2.4745 -0.6437 -0.5716 -2.1401  0.4375 -0.7477  1.1687 -0.0875
#>  0.2272 -2.0428 -0.9572  1.2742  0.1240 -0.6919 -1.7568  0.2886 -0.4121 -0.8140
#>  0.4795 -1.0098 -0.2739 -0.0747 -0.4101 -0.2986  1.5707 -0.2498 -1.1637  1.0247
#>  0.7700 -0.5297  0.2913  1.0759 -0.1665 -0.1352  0.6599 -0.2422  0.0795 -0.7230
#>  1.1881  0.9410 -1.0163 -0.3568  0.8702  0.4412 -0.4028  0.6686  0.8897  0.7515
#>  0.6408 -0.4580 -0.1465 -0.3725 -0.5928  0.7858 -0.2388  1.6807  0.0173  0.0812
#>  1.3907 -0.2382  1.3879  0.4500 -0.7365  0.6344 -0.7197 -0.2276  0.6583 -0.3242
#> -0.6292  0.0475 -0.5162 -0.9106  0.0473 -0.3951  0.7989 -0.5612 -0.4630 -0.5159
#> -0.4424 -0.1665 -0.7097  0.4578  1.1230 -0.3817  0.2117  1.2337 -0.7716 -1.3094
#>  1.1648 -0.0592  0.9751  1.8241 -1.2357 -0.6445  0.3743  0.3619 -0.5506 -1.3611
#>  0.6688 -1.2066  1.5568 -0.1944 -1.9882 -1.7433  0.6563  0.0101  1.9796 -0.0055
#>  0.4695  0.7396  1.1679 -0.8333 -0.1426  0.4654  0.2310  1.6504  0.3709  0.2856
#>  0.5270  2.0758  0.4923 -0.3927 -0.7500  0.3515 -0.4069 -1.0285  1.8309  2.9426
#> -1.5848  1.0502  0.7698 -0.3533 -0.5156 -0.4773  1.2384  0.2011 -1.4003  0.7203
#> -1.2816  0.4025  0.0761 -0.5399 -0.8837 -0.1842  1.2749  0.8064 -0.7466 -1.4910
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]