Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.8283  0.2282  0.4217 -2.1557 -0.0938  2.0842  0.1078 -0.2367 -0.0894 -0.1687
#>  0.0080 -1.5022 -0.0617 -0.4574 -0.5668  0.7821  0.3457  0.8016  1.1293 -0.0577
#> -0.4136 -1.6619 -1.0287  0.0823  0.7372  0.1476 -0.5115 -0.6095 -0.5970  0.5228
#> -0.6330 -0.0924  0.4134 -0.6976  0.8930  1.4077 -1.0486  1.7425 -1.9238 -0.5675
#> -1.6923  0.6226 -0.4848 -0.0761  0.8587  1.0828  1.5456  0.2364  0.0775 -0.5571
#>  1.4412  0.4894  0.4783  1.1969  0.1079 -0.9334 -1.7680  0.8511  0.9255 -0.3125
#>  0.3260 -0.2924 -0.3154 -1.1962 -0.6552 -0.6057  0.4143 -1.3383  0.5305  1.2560
#> -1.0453 -0.1310  0.9966  0.3878 -0.2260 -0.1695 -2.7347  0.0803  1.4763  0.3857
#> -1.1913  1.5763 -0.4149 -1.0692  0.0115  0.8497 -0.5431  0.0581  0.5533 -0.2935
#> -0.6407 -0.5575  0.4743  1.4675  3.1649 -0.5259 -0.8980 -1.4464 -1.1378 -1.4015
#>  0.0298  1.3338 -0.4432 -0.4567 -1.6569 -0.3014  0.2659 -1.0970 -0.2152  2.7150
#> -0.3561 -1.0679 -1.1046 -0.7092 -0.3343 -0.5450  0.6782 -1.1720 -0.5836  0.7750
#> -2.1749 -0.9439 -0.2341 -0.6124 -0.6357  0.7659 -0.3193  0.4498 -0.0040  1.1575
#>  0.7117  0.3453  0.0216  0.1174  0.2334  0.1875 -1.8159  1.0822 -0.9072 -1.1449
#>  1.4724 -2.4005  0.8461 -0.5370 -1.3519 -1.0588 -0.7509  0.6161 -0.2331 -0.2193
#> -1.0026  0.2633 -0.8603  0.6967 -0.0892 -1.0633  0.6210 -0.5107 -1.0435  2.4251
#>  0.6304  0.8087 -0.5550  0.2534  1.4885  1.6500 -0.0038 -0.0768  1.8197  0.5141
#>  1.4723 -1.1361  1.5741  0.9036 -0.8562  1.7764  0.2516  1.5369  0.8250  0.4255
#>  0.1496  0.3454  0.7674 -0.6587  1.2451  1.0822 -1.1224 -0.2183 -0.7550  1.1438
#>  1.2253  0.1814  1.8055  2.1762  0.8547 -1.0186  0.0834 -0.2398  0.0545  0.3314
#> -0.8706  0.2068 -0.1059  1.5114  0.3749 -1.0593  0.3568  0.4006  0.9231  0.8449
#> -1.6893 -0.9138 -2.7286  2.0080  1.1881  1.5330 -0.1284  0.4564 -1.4234  1.0465
#> -0.4592 -0.5362 -0.4189  0.5702 -1.4319 -0.1478 -0.6349  1.0513 -1.4268 -0.4710
#> -0.5236  0.1780 -0.2553  0.4006 -3.1490 -2.0772 -0.6339  0.3285  1.1995 -1.0877
#>  0.2608 -0.9606 -1.4613 -1.1751  1.0310 -0.0425  0.9164 -0.7282  0.1835  0.6442
#>  1.3762 -0.3732  0.3801 -1.3579  0.0139  0.4530 -0.4084 -0.2067 -0.7612  2.6159
#> -1.5366 -0.0561 -0.1667  0.9019 -1.7279 -0.7241 -1.3281  1.1038 -1.6010  0.1066
#>  0.5471  0.9123 -1.5143 -0.0036  0.6560  0.4250  0.8281  0.6362  2.2478  0.3258
#> -0.5256  1.2880  0.2044  0.3242 -0.5948  1.4370  1.3195 -2.0389  0.5737 -1.0833
#>  0.7465 -1.2545  0.8665  2.2089  0.9689 -0.8396  1.9022 -1.3326  0.7517 -0.8302
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]