Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.8703 -0.4812 -0.7476  0.3416  0.6738 -1.4034  0.4066  0.2836 -0.3469  0.0611
#>  0.3309  0.1842  0.1464 -0.1872 -0.3171  0.0675  1.5237 -0.9339  0.9464  2.5155
#>  0.8316 -0.9600  0.5488 -1.2897 -0.3574  1.0441  2.2007 -0.3730 -1.7257  0.8444
#> -1.6352  0.8780 -1.6282 -0.7356  0.5574  0.4228  0.1212  0.0523 -2.5788 -0.2701
#>  0.6172  1.0761  2.2195 -0.4924 -1.1569  0.1553 -1.2915  0.6618  1.4165  1.4724
#>  1.0583  0.3823 -0.2398 -0.1266  0.7470 -0.3764  1.3225 -0.1403  0.1245 -1.4469
#> -1.0890  0.6751 -0.2183 -0.3672 -0.1201 -0.9590 -0.7196  1.0778 -0.8808  0.4435
#>  0.2966  0.5803  1.3299  0.6445  0.4358  0.1508 -0.7617 -0.3686 -0.4431  1.2303
#> -1.1733 -0.5349  0.0450 -0.4946  1.3804  0.0258 -2.4120  0.4911  0.0652  0.2117
#> -1.5614 -0.4895  0.0233  1.3371 -1.9177  2.7599  0.1012 -2.1476 -1.4142  0.2671
#>  0.7683 -1.3114  0.7986 -1.3683  1.1334  0.0223  0.9632  0.2334 -0.9792 -0.9514
#>  0.6634 -1.0279  2.2432 -0.1315  1.5977 -0.3812  0.2088 -1.1060 -2.1324 -0.6541
#>  1.4440  0.5885  0.1642  1.3644 -1.5511 -1.3252 -0.8796  0.0208  0.2808  0.4088
#>  0.8015 -0.0719 -0.3491 -0.3495  0.1033 -0.4168 -0.7290  0.6494  0.7688  1.6953
#>  0.0707  2.1068 -1.5047  0.4582  0.0603  1.2764  0.3755  0.2647 -2.2204  1.3620
#>  0.0505  0.0248  1.1447 -0.0838  1.0225  0.5471 -0.6093 -0.0737 -0.1968  1.2513
#> -0.3260  0.8132  0.1727  0.1345 -0.6844 -1.2841  0.4208 -0.6265  0.8615 -1.5331
#>  0.2739  0.3140 -0.2360  1.1957 -0.1257 -1.8468 -0.5597 -0.5510  0.8277 -1.6596
#> -0.0174 -1.2190 -0.3961  1.3272 -0.8521  0.0738 -0.4117  0.1340  0.4448 -1.0635
#>  1.0456 -0.7747 -0.7313  0.2238 -0.7760  0.4326 -0.0972 -0.1427 -1.1291  0.0186
#> -0.2432 -0.0441 -0.5212 -0.5066 -0.1344 -1.0681  2.5326 -0.2641 -0.5943 -0.2139
#> -0.5791  0.0786 -0.1667 -0.7801 -0.6196 -1.1324  0.7426  2.2640 -0.2098 -0.2496
#>  1.0787  0.8503 -0.8599  0.7212  0.4254  2.0811  0.3134  0.8686 -0.7535  0.4565
#> -0.4727  1.5470  1.2831 -2.1040 -0.6881  0.5129  0.4367 -0.4408 -1.3689 -0.2805
#>  0.1015  2.0677 -0.8350 -0.4006  0.7006 -0.2673 -0.2010 -0.4695 -0.5723  1.6212
#> -0.0012 -1.8774  0.5977  0.9040  0.5200  0.5271  1.1855 -0.4059 -0.2673 -0.1888
#> -0.2221 -0.9674 -0.0169  0.5801  0.0018 -0.2636  0.1112 -0.7963 -0.2077 -0.4322
#> -0.6065 -0.9450 -0.1211 -1.9316 -0.7638 -1.6539  0.3646  0.6378 -2.0486  0.4006
#>  0.0194  0.5595  0.3669 -0.9803  1.1415 -0.9696 -0.9727  0.2965  0.3279  0.8020
#>  2.0481  0.0204 -1.3637  0.6170 -0.3205 -0.6379  0.7020  1.4768  1.3378  0.1776
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]