Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.4393 -0.5537 -0.2167  1.5840 -0.6773 -0.2497  0.2317  0.3324 -1.7029 -1.1629
#> -0.9579  0.7336 -0.7437  0.6283  0.0749  0.1214  0.2058 -1.2519 -1.1465 -1.1717
#> -0.9495 -1.1211  0.0313  2.1878 -1.0691  2.2918  0.1220 -0.6268 -1.2643  1.2975
#> -0.8728 -0.3515  0.0557  1.0157  1.2204  0.0532 -0.1796  0.0153 -0.6200  0.1798
#> -0.9256 -1.6189 -1.2504  1.2381 -0.4034  0.5367 -1.8020  0.6571  1.6967 -0.8695
#> -1.7211 -0.1470 -0.1001 -0.3882 -1.3856  1.0942  1.2357 -1.0500 -0.8138 -2.2401
#>  1.7860  1.5868 -0.5592 -1.7021 -0.2442 -1.8876  0.6680 -0.1737 -0.0468 -0.7088
#> -0.0700 -0.4192 -0.6338  1.1459  1.0351  0.4905 -1.1625  0.8907  0.4682 -0.0256
#>  0.7786  0.5791 -0.9753  1.7066 -0.9590 -0.7450  1.8649 -0.5299 -0.5307  0.2058
#>  0.7379  2.0304 -0.7490  1.1107  1.3965 -0.3858  0.1263  0.6881 -0.2704  0.1656
#>  1.9795 -0.0088 -0.3350 -0.3186 -0.5226  0.2549  0.5502 -1.3486 -1.1797 -1.4091
#> -1.3547 -0.2577  1.3310  1.0597 -1.5566  0.6188  0.8408  0.4908  1.1917  0.0008
#>  1.3065 -0.1945  1.2704 -0.0267 -1.2832  0.8349 -0.5257 -0.0415  0.1595  1.6673
#>  1.2614 -0.1302  0.2873 -0.4030  0.1100 -1.6645 -0.4671 -0.7352 -0.3669  0.3984
#> -0.5243  0.5734 -2.2050  0.6101 -0.4372  1.3278  0.2444 -0.0659  1.4579  1.3070
#> -0.0170 -0.9181 -2.0676  0.6550  1.2736  1.3059 -0.7893  0.3281  0.0025 -0.9941
#>  0.4702 -1.8180 -1.1187  1.0326 -0.1840  2.0832  0.4469 -0.4630  1.4130  0.2712
#> -0.9833 -1.3449 -0.7041  0.5590 -0.1035 -1.1780  0.7551  0.7122 -0.2775 -0.7086
#> -1.4591 -0.2701 -0.1176  1.5177 -0.5313 -1.4089  0.6505 -1.9223 -1.5076  0.3452
#> -0.1344  0.3636 -0.9071  0.1813  0.3749 -1.2287  1.6534  0.4392  0.7265  0.9218
#> -0.3724 -0.6392 -0.6655 -0.3410 -1.5015 -1.0774 -1.7407 -0.7347  1.2535  0.2370
#> -0.7108  1.0927 -0.6937  1.0766 -1.1614 -0.3210 -0.6108  0.8450 -0.5703 -1.1349
#>  0.2202  1.1024  1.6427 -1.7040 -0.4868 -0.6477  0.2727  0.9598 -0.2092 -0.7323
#>  0.4583  0.6007  0.8409  0.7743 -1.2975  0.0477 -0.2967 -1.4443 -0.9649  2.1436
#> -1.6853 -0.4286  0.1414  0.7514  0.0258  0.5606 -1.0156 -0.7703  1.2237 -0.1003
#>  0.5696  1.0077  1.0382 -0.4736  0.7455  0.6064 -0.9023 -0.8584 -0.2891  0.2771
#> -1.2234  0.2211  0.7422  0.2515  0.5131  1.1302  2.9044  1.9308  0.1675  0.1610
#> -1.1798  1.3025 -0.4438 -0.9506 -1.6326 -0.0121 -0.8530  0.6391  0.7125  0.1698
#> -0.7646 -0.4152 -1.6703 -0.7530 -0.4499  0.6122  0.8852  1.3532  0.1340 -0.9606
#>  0.2358  0.7488 -0.8270 -1.1410  1.6221 -1.1484 -1.6982  2.5564 -2.4837 -0.5591
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]