Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.3037 -0.1980  1.0436 -0.4981 -0.5280  0.2282 -0.5688 -0.6815  0.4638  0.2157
#> -0.7319 -1.2013  0.3605 -0.9388 -0.7013  0.3480  0.6026 -0.4705  2.0504 -0.2594
#> -2.5620 -0.6529 -1.0650 -2.1493  0.5922  0.6380  2.0597 -1.0309 -0.2029 -0.1000
#>  0.8189  0.3474  1.0647  0.3067  0.4705  0.4429  0.0470  0.6776  0.6405 -0.2266
#>  0.0270 -1.0669  0.1714 -0.4583 -1.8007  0.7298  0.5536  0.1750  0.1263 -0.1208
#>  0.7359 -0.1099 -0.6567  0.6695 -0.7754  1.2872 -0.2695 -0.9594  0.9649 -1.2119
#> -0.0570  0.0632  0.2832 -0.3945 -1.5677  0.3920 -0.6472  1.3206 -1.6116  0.1366
#> -0.3045 -0.3090  0.5265 -0.6970  0.3241  0.1968  1.0510  0.0462  0.6260  1.3525
#>  0.9149 -0.1048 -0.1269 -0.6420  1.4176  0.3190 -0.2991  1.2299  1.6602  1.0864
#> -1.3599  0.1042 -1.1316  0.1820 -0.1524  0.0354 -1.9181  0.3414 -1.6727  1.6558
#> -0.1508  1.7451 -0.7417 -0.1165  1.0875  0.0952 -1.1463 -0.2142  1.1973  0.5286
#>  0.0237 -1.2086  0.2700 -0.8562  1.0339 -0.0912  1.2272  0.3923  1.3058 -0.3410
#>  1.0569  0.8365  0.5701 -0.9324 -0.0096  1.4427  0.0227  0.3181 -0.8798  0.7737
#> -0.4623 -0.7417 -0.2481 -0.1141  1.2834  1.2412  0.5789 -1.2801  0.2063 -0.6698
#> -0.6232 -1.2402 -1.7319  0.0439 -0.3474  0.1446  0.2214  0.2858  1.7769 -0.0795
#> -0.7844 -0.0714 -1.6322  0.7714 -0.3455 -0.2779 -0.2047  0.3965  1.1444  0.5850
#> -0.8558 -0.1646  0.6439 -0.6352 -0.4884  0.7750  1.1007 -0.7093  0.9343 -0.8273
#> -0.6989  0.7851  0.1360 -0.1400  0.0564  1.6218  0.2810  0.4322  0.1668  1.0039
#>  1.1648  0.0720 -0.3755  1.4968 -0.1033  0.6738  0.4874  0.1731 -0.9297  1.0902
#>  1.1652 -0.1208  1.3366  0.3272  1.3167 -0.5498  0.6562 -0.4762  0.1197  0.8268
#> -0.2970 -0.9969 -1.2511  0.8225  0.3085 -1.3816  0.1361 -0.2303 -0.4223  0.4479
#> -0.5708  0.3070  2.1036  0.6350 -0.3858  0.1696 -0.4408 -0.1228 -0.6189 -1.6764
#>  0.7659  1.0814  0.8403 -0.3450  0.2497  0.8184  1.0788  0.1099  0.6819  0.4958
#>  1.5348 -0.5172 -0.2838  0.0671 -0.4788  0.2751  1.0291 -0.8558 -0.8467 -0.4207
#> -0.1512 -2.0706 -1.1806 -0.1839  0.2321  0.3532  0.7089  0.0203  1.3204 -0.2563
#>  0.1172  0.5234 -2.2758 -0.1104 -1.1584  0.1255 -0.8052  1.9003  0.3512  1.4812
#>  1.0138  1.0449  1.1950  0.1535 -1.1787 -0.5314  1.3499  0.4931 -0.7096 -0.3640
#> -0.8144  2.4281 -0.9782  0.0996 -0.4276  0.6683  0.7659 -1.2093 -0.7991 -1.0588
#> -0.6295  0.2919  1.2815 -0.7167  0.3426  0.1205  0.7620 -2.5135 -0.1836  2.6154
#>  0.2223 -2.1197  0.5851 -1.2075 -0.7138 -0.4181 -1.1606 -0.7537 -2.5301 -1.2434
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]