Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.3288 -0.9449  0.8422 -1.0211 -0.2158 -1.0152  0.7875  1.6588  1.2028  0.1283
#> -2.2005  0.6886  1.0489 -1.4330  0.1800 -0.4567  0.2186  0.3602 -1.8095  0.0636
#> -0.4753  0.2929  0.0420 -0.0964 -0.2653  1.1039 -0.0827 -0.0934 -1.4090 -0.4278
#>  0.8414 -0.7602  1.7192 -0.3978  0.4415  1.6289  0.1150 -0.3149 -1.4337  0.2220
#>  1.5213  0.5500  0.1393  0.6359 -0.1464 -0.4334  2.7792  0.9999  0.8161  0.6482
#> -0.8713 -0.1758  1.5472  0.5723 -0.9805 -1.3486 -0.5539  2.3291  0.7830 -0.7237
#>  0.7034  1.1641 -0.2035 -1.2067  1.1367  1.4298 -0.7124 -0.1010  1.8075  1.7226
#> -1.3624  1.5134 -1.1745 -0.5225 -0.3529 -0.3058 -0.3303 -1.2481  1.4699 -0.6093
#> -0.1370 -1.2792  0.0722  0.2598  1.3793  1.8070  1.0128  0.2605  0.7657  0.1271
#>  0.9765  0.2173  0.3143  0.4061 -0.8170  0.2531 -0.3085 -1.3325  0.5234 -2.4569
#>  1.0191 -0.0364 -0.2674 -0.3190 -0.3497 -0.0845 -0.1522 -0.0811  1.7401 -3.3245
#>  0.0783  0.5342  0.4711  1.0173  1.2655  0.1982 -1.6773  0.2932 -0.5840 -0.1103
#>  1.1820 -0.0287  0.5281  0.6912  0.4900 -0.1418 -1.1483  0.1929 -0.0407 -0.6680
#>  1.2654  0.0596 -0.9034 -1.3467  1.6951  0.1824  0.9054  0.5405 -0.9530  0.3293
#> -0.8221  0.0660 -0.6172  1.7676 -0.3520  0.3183  0.1765  0.4497  0.4503 -0.9394
#> -0.4063 -0.8390 -1.7008  0.9373 -1.8218  1.0392  0.2537  0.1964  0.1182  0.5303
#> -0.1746  0.6685 -2.0158 -1.1044  0.0041 -1.0498 -0.7444  0.7024  0.5287  0.0987
#> -2.4539  0.4338  0.0040  0.8964  0.3416 -0.5326  0.3257 -0.8669  1.7761  0.0327
#>  0.8004  1.4894 -1.0811 -0.9946 -0.8640 -0.2245 -0.6210 -0.7302 -1.3680  0.5521
#>  0.1573  1.1023 -0.2096 -0.4072 -1.9126 -0.3848 -0.7351  0.6319  1.1514  0.3811
#> -0.7192 -0.1659  0.3347 -1.5329 -0.1038 -1.4061 -1.0248 -0.2231  1.0395 -0.2476
#>  1.6730  0.2818 -1.6099 -0.3088  0.0535  0.2281  0.7980 -1.0849  0.7577 -0.5070
#> -0.7785  1.2866  0.8771 -1.7292 -1.4055 -0.5439  1.7700  1.5387  1.5714 -1.2262
#>  2.6137  0.8274 -2.0244 -0.9606 -0.0071 -1.5194  0.4798 -0.2315  0.1206 -0.3949
#> -1.8934  0.8163 -0.4129  1.7590 -0.5858 -0.5462  0.0519 -0.6316 -1.4664  1.7480
#>  0.4157  0.2144  0.7667  0.2877 -0.5648  0.6777 -0.3417  0.5269 -1.5164 -0.2760
#> -1.5254 -0.0244  2.8521 -0.9821  0.3529 -1.4257  0.7976 -0.5835 -0.3050  0.6138
#> -0.8540 -0.5568 -1.7018 -0.5748  2.6277  0.9434  0.9089 -0.0738  0.2777 -1.0564
#>  0.5130  0.1703  0.6400 -0.6900 -1.1061 -0.6394 -0.7364  0.2872  2.3283 -0.4388
#>  1.5346 -0.3760  0.0961 -0.9274  0.5336  0.5350  0.0746  0.2387  0.6167 -0.2766
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]