Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.6674 -0.2872  0.1388 -0.5804  1.9945 -0.2273  0.6087 -1.0733 -1.0983 -0.4383
#> -0.4313 -1.0561 -0.8578  0.2119 -0.0952 -0.0448  0.1990 -0.5197 -0.3429  0.9948
#>  2.9620  0.0827  0.6897  0.6871  1.0383 -0.3161 -2.6809 -0.7953  0.8829 -1.2481
#> -0.9257 -0.8043 -1.1097 -0.2678 -0.0713 -0.5581  0.3611 -1.0673 -0.8742 -1.2988
#> -0.4791  0.2766 -0.3004 -1.5015 -0.2575 -1.0186 -0.1912  1.0592 -0.4771  1.3646
#> -0.0129 -0.4482  0.5867 -0.0088 -1.0360 -0.5368 -0.8208  0.2315 -0.5089  0.2980
#>  0.3291  0.4099  0.0589  0.2568  0.8847 -0.6065  0.3656  0.4765  0.9134  2.9770
#> -1.5495 -0.6799  0.3703  1.5399 -0.4931 -0.3949  1.3124 -1.2803  0.4899 -0.6581
#>  0.5167  0.4670 -1.3304  0.0195 -1.4518  1.1452  0.9159  0.3314 -0.0115 -1.0417
#> -1.7550 -0.2913 -1.1888  0.1612 -0.1810 -0.0879 -0.7966  1.1129 -0.3054  0.8201
#>  0.6261 -0.0465 -1.1766  0.5829 -1.8018  1.3772  1.7959 -0.3234 -0.4249 -0.2674
#>  1.5942  0.6726  2.2215  0.7697 -2.3233  0.7298  2.8574 -0.5104 -1.6248  1.8346
#> -0.1754  1.1547  0.1112  0.5659 -0.9992  0.6938  1.6175  0.9239 -1.3837 -0.6761
#> -1.1921  0.0600  0.5866  0.0941 -2.0214  1.3340  0.4103 -1.3896  0.2604  0.5112
#>  1.1235  1.2360 -0.9786  1.2145  0.1616  0.2475  1.3070  0.4679 -1.2935 -0.6551
#>  1.0404 -0.3665 -0.2713  0.2168  0.1291 -0.2478 -0.7138 -0.4828 -0.6293 -0.9939
#>  0.6834 -0.5597  0.0495  0.5237 -1.3754  0.2552 -0.1253  0.2196 -0.0871 -1.1483
#> -1.8514 -0.5344  0.8105  0.6650 -0.5803 -0.7044  0.3501 -0.1659  1.7420 -1.8915
#>  0.2934 -0.6626 -0.7159  1.2906 -0.5403  0.0231  0.7513 -0.3725 -0.6637  0.5495
#> -0.3244 -0.3896 -0.8227 -0.2080 -0.1719  0.3390  0.7407  0.4920 -0.5399 -0.2687
#>  0.1622  2.3405 -0.4696  1.7715  0.9647 -0.3139  0.5411  0.8753 -0.2923 -0.0746
#> -1.6088  0.9285  0.8963  0.5365 -0.6624 -0.2992  1.1234  1.4924 -0.2991 -0.2848
#>  0.3769  0.2892 -0.2750 -0.6144 -0.5683  0.8001 -1.1687  0.9902 -0.6907 -0.5573
#> -1.3845 -0.6464 -0.5433  1.3356  0.9650 -0.2449  1.9431  0.4565 -0.3043 -0.0775
#> -0.7875  0.5611  1.5078  1.1709  0.1660  0.2279 -1.4189  1.0878  3.0059  1.8947
#>  0.0914  0.2730  0.4892  0.4495  0.2886 -1.0115 -0.7591 -0.3774 -0.0412  0.8887
#> -0.5720  0.4326  1.1983  0.5122 -0.8514  0.3165  0.7052  0.3789 -0.3425  0.3531
#> -1.3818 -1.0354 -1.4471 -0.0848  0.9038  0.0099 -1.2402 -0.7003  1.0202  1.5493
#>  0.4026 -0.2989 -0.2056 -0.7119  1.2114  0.2706  0.4437 -0.8622 -1.8651  0.5882
#> -0.3113  0.7463  0.0277 -0.1660 -0.4855  0.5812  0.5052 -1.3882 -0.2610 -0.0559
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]