Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.1527  0.0378  0.6281  0.6516 -0.1145 -2.2069 -0.2797 -0.6174 -1.5739 -0.2476
#> -1.5667  1.0743  0.3640  0.9652 -0.0222  0.3653 -0.6498  0.3952 -0.2389  0.0604
#>  0.8755  0.9450  1.2145  0.9435 -0.3804  1.5239 -0.3292 -0.6880 -2.6503 -1.4679
#> -0.1509 -0.3152  0.2412 -2.4644 -0.3944 -0.3164  1.8330 -0.4522  0.5373  0.6468
#>  1.4067  0.4406  1.4156  1.2681 -1.1447 -0.4386  0.2391  1.0646 -1.3884  1.4255
#>  1.0420 -0.4923  0.7501 -0.4194  0.1152 -1.2423  1.1067  0.5215 -0.4559  1.5979
#> -0.5972 -1.2053  0.6053  0.3282 -0.2771  0.3292  1.2041  0.4042  0.2265  0.3517
#> -0.1314 -0.6943 -0.1520 -1.5789  0.7236  0.2902  2.2691  0.3769  0.2311  0.0166
#>  0.9106 -0.0449  0.8894 -0.7175  0.1708 -2.1904  0.7771 -1.4192 -1.2582  0.2467
#>  0.8944  0.5320  0.5023  1.0301  0.4542 -0.1338 -0.4640 -0.6514 -2.0486  0.2438
#> -2.0520  0.6981  0.0732 -0.0027  0.1863  0.3443 -1.9216  0.0784 -2.0891 -0.5362
#> -0.9176 -1.7318  1.0836  2.3848  0.7454  0.8974  0.8457  0.0980  0.8483 -0.2969
#> -0.2219 -0.9921  0.0041  0.4066 -1.3185  2.7540 -1.6570  1.1202  1.7455 -0.5784
#>  1.2910 -1.6417  0.2656 -0.4958  1.8911 -0.9619 -0.3184  1.9242 -0.3324  0.1815
#> -0.7544 -0.5503  0.6650 -0.8182 -0.3221  0.0063 -1.2838 -0.2115  0.2102  0.8688
#> -0.8125 -0.2187 -1.4056 -0.1426  0.3774  1.3148  0.7119  0.6565 -2.3078 -0.7171
#> -0.3392 -0.4923 -1.9075  0.9102  0.8049  2.2752 -1.3361 -0.5580 -0.9541 -0.1998
#>  1.8748 -0.6449  0.2045  1.1586  2.2635  1.6711 -0.6562  0.0363  0.8977  1.4444
#> -1.4592 -1.0255 -0.0185  0.1270 -1.5497  0.1724  1.6783 -0.1368 -0.0910  1.1014
#>  0.1908  0.9750  1.0960 -1.2603  1.4856 -1.0637  1.4551  0.9435 -0.5846 -1.0617
#> -0.2299  0.3083  1.3438  0.4289  0.5616 -1.6534 -0.5402  0.2003 -1.7066  0.2464
#> -0.9631  1.6652  0.7675 -0.1334 -1.4588 -1.4006  0.7915 -1.1032 -2.4208  0.0610
#> -1.4195 -0.5984  0.4870 -0.4912 -0.3727 -0.1657  0.8944 -0.1461 -0.7361  0.2358
#>  0.8607  0.5487  1.2833  0.7455  0.4491  1.1908  0.2985 -0.5532  0.2733 -1.0351
#> -0.9489 -0.3232  0.3908  0.7965  0.5958  0.1822  1.1076  0.7950 -0.8873 -2.0241
#>  0.5643  1.4937  0.0210 -1.5689 -0.6031 -0.6696  1.0210 -0.1158  0.8737 -0.1653
#>  0.2573 -0.0164  0.5035  0.6214  0.0749 -0.3992 -0.4251  0.1253 -0.8700  0.6454
#> -1.2053  0.0464 -1.3263  0.7116 -0.0998  0.6180 -1.0124 -0.5683 -0.6565 -1.1629
#> -0.8339 -1.3435  1.3817  0.5943 -1.1110  0.4687  1.5485  1.1898  0.9648  3.5279
#>  0.6733 -1.3047 -0.5843  0.7157 -0.8817  0.6603 -0.3479  0.2590  0.3008 -0.9696
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]