Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-2.1009 -0.0325  0.8200  0.3445 -1.5187 -0.8997  0.2894  0.8026 -0.1466 -1.6486
#>  0.0873  0.7565 -0.2600  0.2954 -2.7724 -1.4893 -1.0389  1.3610  1.1459  0.2038
#> -0.8819 -0.0974 -0.3500 -1.1982  2.3434  0.2886  0.8671 -0.2203  1.1956  0.6572
#> -0.5870  2.5842 -0.2525  2.3126  0.8205  0.3730 -0.2647  0.2152  0.4430 -1.4450
#>  1.5654 -0.0435  0.0359 -0.2176 -1.3248 -1.3274  1.5261 -1.1994  0.2893  1.1235
#> -0.3992 -1.6315 -0.8311 -0.4873 -0.0309 -1.3816  0.1634  0.1266 -0.3739 -0.4819
#> -0.0116 -0.3940 -0.6367 -1.8769  0.6779 -1.3798  0.2673  0.5780 -1.1814  0.3880
#>  1.0374 -0.1765 -0.5554  0.3781 -2.0245  0.0099  0.2049  0.9164 -0.1084  0.6505
#> -1.0988  0.3285  1.4350 -1.0479 -1.5008 -0.1177  0.6284  0.2284  1.5508 -0.2495
#>  0.5228 -1.4227  0.3886 -0.3990 -1.6055  0.5965 -1.2525 -0.3656 -1.3033  0.5136
#>  0.3863  0.5116 -0.2014  0.6820 -0.6909 -0.4692 -1.0523 -0.9473 -0.1871  0.2007
#> -0.5749 -0.3883 -0.0191  1.2325  1.0061  0.2152 -0.8146 -0.1611 -0.2510  2.0400
#> -0.3632 -1.4086 -0.5999  1.6081  1.8965  0.4671 -0.3770 -1.8652  1.2459  0.1840
#>  0.1672  1.8277  1.0873  0.4651  0.0503  0.6138  0.5547 -1.6034 -0.5848  0.5019
#>  0.0859  0.1554  0.5493 -0.2001  1.2253 -1.1635  0.8279  0.3869  0.5557 -1.8030
#>  0.0684 -0.3690  0.2640 -1.0043 -0.8089 -0.8184 -1.0100  1.4817  0.7091  0.8986
#> -0.2825 -0.7221 -0.8054  0.7024 -1.8211  0.4661 -1.5996 -1.6588  1.2603  0.3801
#>  0.4645  0.2474  1.5611  0.8496 -0.1295  0.0756  0.5513 -0.4337  1.4361 -1.4222
#>  0.1392 -0.8894 -0.4364  0.2219  0.6804  0.6470  1.4412  0.4601  1.0710  1.2973
#>  0.5624  0.0783  0.7213  0.1497  0.9147 -3.4691  1.0177 -0.4863  1.4484 -2.3611
#> -0.2403 -1.1915  1.3446  0.4700 -1.1311 -0.1114 -0.1775 -1.5382  1.4066  2.5173
#> -0.3345  0.4120  1.2635 -1.0604  1.2275 -0.8765 -0.2456  1.2884  0.2224 -0.5190
#> -0.3650 -1.8631 -0.0331  0.4722 -0.0174 -0.6669 -1.0684 -1.3055  0.1381  0.7492
#> -0.9246  0.9676  0.8093  1.0844  0.6973 -0.2296 -1.4926 -0.6153 -0.1079  1.3876
#>  0.4161  0.0325  0.3338 -0.8110 -0.9523  0.7026  1.1500  1.0987  0.4401  0.8114
#> -2.3630  0.5800  0.2501  1.7664 -0.7926  0.4908  1.0416 -0.2707 -2.5732  0.8335
#>  0.4119  0.8294  1.3002 -1.2993 -0.6719  0.4267  0.5066  0.8796 -0.9789 -0.0002
#> -1.3564  0.4684 -1.2083  0.5887  1.0039 -1.2275  0.0908 -0.5223  2.1151  0.3314
#>  0.6412  1.3354  0.0699 -0.4952  0.9312  0.2129 -0.8085 -1.0135  0.2281 -0.8686
#> -1.1990 -0.0548  1.3399  0.1682 -0.8133 -2.0438 -1.1772  0.5392 -1.2934 -0.4989
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]