Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.3433 -0.0609  0.4881 -1.2285 -0.8351 -0.9934 -0.1221 -1.7757 -0.7349 -0.3387
#>  1.0612 -0.8913 -0.6765  0.4197  1.8914  0.0740 -2.3555 -1.7426  1.2271 -0.0728
#>  0.0391 -0.5000 -1.0186  1.1668  1.1652  0.9088 -0.6956  1.8653  1.9367 -0.4569
#>  0.0282 -0.6281  2.5930 -0.3686  2.2590  2.2246  0.2743 -0.7170 -0.4571  0.7366
#> -0.3891  0.2339  1.0919  0.4466 -0.6502  0.4938  0.1022 -1.4086  0.9518 -0.9406
#> -0.3125 -2.3083  1.0050 -0.8068 -0.7439 -0.1178  0.2432 -1.5750  0.6443 -0.6864
#>  0.0353 -0.7585  0.6966  0.6151  1.2647  0.0995 -0.7198  0.4318 -0.3873 -1.5495
#> -1.0181  0.3275  0.6866 -0.4282  1.6989  0.1665  0.5621 -2.1054 -1.7828  0.3819
#>  0.0551 -0.3543  0.2882  0.5277  1.3767 -0.8151  1.0926  0.9091  0.2584  1.0385
#> -0.5897  1.3125  0.3230  0.1854  0.7931 -0.0576  0.6348  0.6050 -0.3202  0.3641
#> -1.0598 -0.2881  1.1740  0.2772 -1.3719  0.7577 -1.3152 -0.2888 -1.0150 -0.3636
#> -0.2559  1.0260  1.6421 -0.4791 -2.5336  1.0022 -0.1841  2.2948 -1.2241  0.3749
#> -1.6303 -0.5318 -0.5476 -0.0594  1.6599  2.7592  0.3921 -0.6173 -0.1903 -0.8536
#>  0.3001  0.9124  1.0531 -0.5485 -0.4136 -0.9025  0.8490 -2.5002 -0.4325 -0.1610
#>  1.1540  0.9693  0.5977  1.5373  0.3142 -1.4699 -1.4391  0.8134  0.1368  0.4209
#> -0.2645  1.9160 -1.4287 -1.8199 -0.8145  0.8154  1.1278  0.4994 -0.1213 -0.9769
#>  0.4851  2.4882  0.5481 -0.6093  1.4129 -0.7919 -1.0837  0.2369  0.9450  0.1295
#>  1.1460  0.8666  0.4340 -0.8980  0.2966  0.8034 -0.6203 -2.4563 -0.4710  0.8660
#> -0.0621  0.9838  1.6144  1.0541  0.7773 -0.1475  2.2924 -0.9086  0.0036  0.6483
#> -0.3826  0.9699 -0.2717 -0.0466  0.9776 -0.0326 -1.9700 -1.0432 -1.6480  0.4443
#>  0.4554 -0.7307  1.1687  0.4721 -0.9273  0.7997  0.5354  1.0189  0.2388  0.0937
#>  1.5231 -1.1434 -2.5146  0.3653  0.1056 -1.6950 -0.9335  0.0832  1.9238  0.5769
#>  2.0769  0.7463 -1.1118 -0.3571  0.5789 -1.0198 -0.8095 -2.1312 -0.2884 -0.3535
#>  0.2013  0.3416  1.5060  0.4238  1.3194  0.1314  0.0081  0.8035  0.9100 -2.9247
#>  0.9990 -1.2301 -0.5216 -0.8005  0.8082 -1.2250 -0.3435  0.7471 -0.1461  0.0065
#>  0.1176 -0.7711  1.4563 -1.0540 -1.0240 -0.6515  0.3246 -1.4149 -0.3889 -0.1395
#>  0.0174 -0.1082  0.6370  0.1474  0.7082  2.3682 -1.5498  0.3086 -1.4865  0.1923
#> -0.7377 -1.9122  0.9086  0.9426  0.5314 -0.7529  1.7941 -0.7229  1.8202 -1.4007
#>  0.7946 -1.8832 -1.1604  0.5517  0.0852 -0.6863 -0.2452  0.2643  0.3099 -0.4510
#>  0.5149  0.4991 -0.4641 -0.8673 -2.1432  1.6665  0.4922 -2.0546 -0.3676  0.8672
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]