Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.4388 -0.1098 -1.8762  0.3332 -0.9893 -1.6304  0.3293  0.0422  1.0476  1.5244
#> -1.0023  1.9528  1.4199 -0.7194 -0.4292 -1.0805  1.2953  1.1173  0.6679 -0.4472
#>  0.1855  0.0695 -1.1927  1.4276  1.5819  0.8325 -0.5569 -1.2571  0.0024 -0.0360
#>  0.2471 -0.1433  0.0717 -1.2792  0.1170 -0.3690 -0.5680 -0.2838  0.2795 -0.4199
#>  0.2810 -0.3614  0.9594  0.5319  0.0215 -0.4335 -0.7540 -0.6062  0.5929 -0.6866
#>  0.1571 -1.3652  0.3642  0.0944  0.0449 -1.1860  0.5103 -0.4957  0.6913 -0.0146
#>  2.2448  0.2435 -1.1043  0.4431 -0.5374  0.6777 -1.5094 -0.0509 -0.2991 -1.2014
#> -0.5372  0.4614  0.9152  0.8647  2.0941  1.2894 -0.6575 -1.9719  0.0220  0.8199
#>  1.7658  0.3930 -0.3388 -0.1669  0.4208  0.0082  0.5123 -1.6612  0.8316  0.3953
#> -0.6709 -1.3121 -1.2269 -0.4785  0.8045  0.0602 -0.5870  0.8147  0.9547 -2.3983
#>  0.5121 -0.3363 -1.9456  0.7328  0.0757 -0.2324 -1.4422  0.2917 -0.9066 -0.1501
#> -1.3056  0.3136 -1.0260 -1.3553  0.0292 -1.7509  0.0431  0.8446  0.6927  0.5908
#>  0.5071  1.4916 -0.9656  0.0951  0.6937  1.4341 -1.9758 -0.7595  0.2257  0.6641
#>  0.3360  1.1773  1.2276 -0.4437 -1.5500  0.4676 -0.1276  0.0961  0.1079  0.9150
#> -0.4293 -1.1924 -0.8267 -1.3914  0.2793 -0.6993  0.6041 -0.2053 -0.9420  1.5344
#> -0.1627  0.8964  0.1529  0.9259 -1.3797  0.2577 -0.0099 -1.4387  0.1454 -0.6612
#> -1.5404 -1.6405 -0.9638  0.1847  1.5223  0.7545  1.0049 -0.4259 -0.9506 -0.5505
#> -1.1765 -1.2902  1.1442  0.7294 -0.7048 -2.0002 -0.9462  0.3265 -0.7129 -0.9835
#> -1.0261 -0.9233 -0.7720 -3.0662 -0.5934 -1.3684  0.3042  1.5960 -3.6786 -0.1946
#>  0.1552  0.2595 -0.2339 -0.1014 -0.2009 -0.6275 -0.3558 -0.4503 -1.0717  0.7266
#> -0.5951 -1.5013 -1.2048  0.8589  1.6465  1.2249  0.1017  0.8941 -0.6732  0.0863
#> -2.8961 -1.8432 -0.3222 -0.9341  1.1218  0.2142 -0.3705  1.3898  0.5636 -1.4192
#> -1.7213 -0.5130  0.5902 -1.1250  0.7273 -0.5704 -1.9084 -2.0866 -0.3643 -1.1320
#>  0.0645  1.1762  0.8387 -0.3371 -0.1955  1.0917 -0.9061 -1.3370  1.3885  0.7411
#> -0.3217  0.5723  0.6428 -0.5813 -0.1609  0.5991 -0.1523  0.0019 -1.5713  0.1115
#> -0.0465  0.5950  1.1888 -0.7023 -0.4425  1.4195 -0.6250  0.0327  1.0521  1.1881
#>  1.3482 -0.8456  1.1533 -1.2256 -0.2929 -1.6763  0.7947  0.4237 -0.0944 -1.2400
#> -0.5727 -0.6424  1.3091 -1.0960  0.4375 -0.3618  1.1956  0.6320 -0.3390  0.8785
#>  0.3493 -1.6354 -0.8889 -2.1592  1.4041 -0.1900  0.9318  0.0709  0.6547  1.7350
#>  1.8198  1.3800 -1.3797 -0.9561  0.2289  0.7937 -0.3735  0.0450  1.3860  1.9002
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]