Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-2.2354  0.6023  0.8136 -1.2396  0.8178 -0.7162  0.8817 -0.8076  1.8083  2.1874
#>  0.3494  0.8282 -0.8139  0.7956 -1.9190  1.4189  0.9452 -0.1264 -0.2948 -0.7304
#> -0.0998 -2.3179  0.0409  0.0549  2.0042  0.8208  1.4098  0.5146 -2.2232  0.8806
#> -1.8664 -0.7472 -0.5508  1.0437  0.1991 -1.6834 -0.9421 -1.1183 -0.5629  0.2220
#> -0.9183 -0.8986 -1.2676 -0.7903  0.1478  0.5204 -2.0150 -1.8148 -0.1116  0.2395
#> -1.0219 -0.5260  0.3881  0.9788 -0.4285  0.0722  1.5899 -1.0475 -0.0875 -0.9885
#> -1.2700  0.9355 -0.4166  0.4664 -1.4067  1.3845  0.6641  0.7181 -2.6647 -0.6161
#> -0.5367 -1.4433 -0.4774  0.1200  1.2708 -1.3664  0.5682 -0.6909 -0.6703  2.2909
#>  2.0450  0.7487 -0.8213 -0.0200  0.2414 -0.9674  0.5926 -1.5831 -0.9190  0.0811
#>  1.2104 -0.0669 -0.9584 -0.4088  1.1196  0.4341  1.7139 -0.6010  3.0789 -0.5565
#>  0.2747 -1.3312 -1.0584 -0.7273  0.4674 -0.1664 -0.9722 -0.2447  0.3444 -0.3801
#> -1.0198 -0.0944  0.2854  0.6426  1.1978  2.1799  1.0367  0.2904 -1.0963  0.2185
#>  2.5098  0.5017 -0.9480  0.7878 -0.6784 -1.1380  0.4233  0.6866  0.7066 -1.5100
#>  0.1873  0.4120  1.1395  2.1939 -2.4501 -0.7970 -1.4546 -0.6729  1.2384  0.5003
#>  1.5204 -0.7836 -0.0126  0.6959 -0.9258  0.3559 -1.0147 -0.6166  0.7421 -0.3031
#>  0.3189 -0.9559 -0.2528  0.1601 -1.4349  0.3191 -2.0140 -1.0416 -0.8427 -1.6729
#> -2.2322  0.1352 -0.5781 -1.0347  0.4023  1.4545 -1.7148 -0.1564  0.5275  1.3518
#> -0.5346 -0.9767 -1.1891 -0.5128 -2.1186 -0.4252 -1.6425 -0.5517  1.1632  0.9701
#> -0.1215  0.9020  1.0541  1.1421  0.8423 -0.9902 -0.9820  0.5892  0.0353  2.2523
#>  1.4798 -0.2662 -0.7704 -0.0024 -0.9938  0.6572  1.0186 -0.5530 -0.0719  1.0432
#>  0.8230 -1.6857  0.2642  0.7752  0.6216  0.8942  0.7776 -0.9138  0.2154  2.4042
#>  0.6579  0.1531  0.6468  1.0672 -0.8295  0.7528 -0.7233  1.8976 -1.3439  0.5779
#> -1.3736  1.4481  0.3960  2.0946  0.0220 -1.0581 -1.0009 -0.4195  0.6272 -0.3580
#>  1.3939  1.1384  2.1546 -0.4326  0.6093  0.1499 -2.5970  0.1521  0.6690 -1.4075
#> -0.7577  1.0036  1.4802  1.7567 -0.6025 -0.1980 -1.5554  0.5676  0.1312  1.6311
#>  0.6476  0.2516 -1.6801 -0.0354 -1.9261 -1.1536 -0.1137  0.1910  1.1893  1.0137
#>  1.3101 -0.3448 -0.0474  0.2752  1.4924  1.8173  0.1383  1.7799  0.2568  0.7203
#> -0.4973  0.2753 -1.5510 -0.0739 -1.5060  1.6403  0.2361 -2.3432 -0.6114 -0.6977
#> -1.1599  1.0105  0.5902  1.3212  0.4401  0.4752  0.3107 -1.2131 -1.6632 -1.3773
#>  0.8555 -0.6985 -0.0517 -0.4433 -1.3301  1.2713  0.7011 -0.4518 -0.9690 -0.8562
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]