Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.8206 -1.0185 -0.7714 -0.1861 -0.1631  0.0319 -1.0711 -0.8500 -1.1806 -1.0063
#>  0.3067  0.7634  0.6058 -1.3014  0.0461 -0.3798  0.9430 -0.1076 -0.0149  0.9269
#> -0.1776  0.1557 -0.7797 -1.7588 -0.1279 -0.8647  0.1568 -0.2124 -0.7317 -0.1021
#>  1.3138 -1.0865  0.5504  1.2637  0.4605 -0.4034  0.3096  0.6336  1.0951 -0.5196
#> -1.1728  1.1997 -1.4744  0.4197 -0.5885 -0.4215  0.3829 -0.7150  0.4098 -0.1784
#> -0.0115 -2.4828 -1.1774 -1.8674 -1.0971  0.2405  1.7875 -2.7310 -0.2656  1.7439
#> -0.3224  0.2364 -0.8296  1.3941  0.5282 -0.5429 -0.7395 -0.6773  0.2398  0.9789
#>  0.7434 -0.2587  0.3704  0.3066 -0.0461  1.4205  0.4317 -0.5098 -0.5039  1.0304
#> -0.0587 -0.1920  0.4396  0.4097  1.4288  0.9935  0.1694  0.3263  1.9905  0.3126
#> -0.9220 -0.8109  0.1107 -0.3485  0.1689  0.1526  0.5696 -0.9781 -0.0975 -0.2809
#> -0.7748  0.9250  1.0531 -0.2578  0.3891  1.6142 -0.6978 -0.5725  1.8693 -1.2433
#> -0.9968 -0.8154 -0.1169 -1.8823  0.1151 -0.5223  0.5714  1.0632 -1.1113 -1.1727
#> -2.0901  0.6717  0.4658 -0.9546  0.3351  1.2169  1.6360  0.0360  0.1679 -1.9012
#> -0.4277  1.1952 -1.0115 -0.3037  0.5602  1.4558 -1.2336 -1.2446 -1.4572 -0.4637
#>  0.2596 -0.1155  0.9869  1.8656 -0.5566 -0.4822 -1.8531  1.4224  0.8065  0.2934
#> -0.4263  1.4743  0.3687  1.6311  2.0613  0.2859 -0.8395  0.9027  0.0825  0.6602
#> -0.9827 -0.5341 -0.6503 -0.3974 -0.8928  1.1414  1.7088  0.8961  1.5450  0.5300
#>  0.4512  0.7766 -1.2151  0.4879 -0.1527 -0.8585  0.2984 -0.3954 -0.5195  0.9029
#>  0.4462  0.6275  2.4105 -1.4732 -0.6169 -0.2101 -0.3437 -0.1481  0.2444 -0.7428
#> -0.4394 -2.4383 -1.7569 -0.2275  1.1980  0.5985 -0.2840  1.3805  1.6218  1.7333
#> -0.3290  0.7183  1.2178 -0.6711  0.8238 -0.7907  0.1801  1.0136  0.4493 -1.4468
#> -0.6684  0.0174 -0.8868 -0.5837  0.0968 -0.2395  0.5636 -0.4646 -0.5370 -0.6543
#> -1.0427  2.0882 -2.9310 -0.3425  0.4270 -1.4498 -0.3723  0.1358 -1.0841  0.6383
#>  1.0570 -1.2918  1.0663 -0.1246  0.3471  1.0581 -0.9473  0.6718  0.2648 -0.8789
#> -0.8493  0.9483  0.2501 -0.7519 -0.2378 -0.2975 -0.0490 -1.8152 -1.7251 -1.7608
#>  0.5790 -0.8080  1.1726  0.3154  1.2151 -2.0256 -0.5727  0.0583  0.6084  1.1040
#> -1.8685  0.3324  0.2355  0.0999  0.9419 -0.3693  0.7634 -0.4504  0.7657  0.5178
#>  0.5656  1.0067 -0.1062 -0.5500 -0.4615 -1.0980  1.1587  0.7500  0.7459 -0.0311
#> -0.6552  1.1523 -0.2282 -0.2912  0.7332 -2.0004  0.2425  0.0879 -0.3188 -0.3731
#>  0.5945 -0.4816 -0.4476  1.2126  0.7673  1.2829 -0.6223  0.4854 -1.2598  1.4154
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]