Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.4443  1.2284  0.1470  0.2033 -0.6888 -0.8215 -0.5909 -1.0523  1.6955 -0.3199
#>  2.2180  2.2491 -0.7675 -0.0063  0.1579 -0.4867 -1.6438  2.4069  0.6342 -0.8884
#> -0.5609 -0.5010  1.6063  1.0356 -0.5153  1.0236  0.5229  1.3255  1.0367  1.0123
#> -0.5012 -1.5117  1.2298 -0.5501 -0.3167  0.5099 -0.6606  0.4801  0.1420  0.1638
#> -0.8148 -0.8910 -1.1717  0.8495  0.9322  0.9468  0.6831  0.7081 -1.2717 -1.0212
#> -1.3420 -0.8550 -1.3735 -0.1425  0.0199 -0.1584  0.2983 -0.2926  0.2112  0.8192
#>  1.0119  1.1131  0.3071 -0.1995  0.8290  0.0993  0.1775  2.3332 -0.0905 -1.5425
#> -0.5527 -0.0108  1.1926  0.2880  0.2593  1.4643 -0.1573  0.0883 -0.1759  0.7940
#> -0.9337  0.2347  0.1427  1.7058  1.1741  1.5157  0.0349 -0.3655  0.7798  0.7099
#> -0.3166  0.2418 -0.5487 -0.0918  0.9786  0.3479 -0.9549 -1.0318  0.6837  1.5044
#> -2.7292 -0.2665  0.0516 -1.4459  0.4325 -0.9537  0.5745 -0.2290  0.5720  1.2124
#> -0.1548 -1.0982 -0.7685  1.9247  1.2980 -0.2342  0.5183  0.0979  2.4397 -1.0259
#>  0.2184 -1.2098  0.3047 -0.2993  1.5001 -0.0248  0.4251 -0.3136  1.5475 -0.7910
#>  0.8103 -0.6932  0.9086 -0.4161 -0.1388  1.0437 -0.4563  0.2431  0.2674 -1.2190
#>  1.2141  0.3649 -2.3340  0.2442 -0.2530 -0.0186  0.6832  0.1041  0.7239 -2.0322
#>  0.0679 -2.3343 -0.5367  0.6350  1.4398 -0.3275  1.5309  0.1705 -0.5230 -0.4514
#>  0.9430  0.7875 -0.5926  0.2083 -0.7211  0.2171  0.0724  0.4902  0.8114 -0.0245
#>  0.5270 -2.5448 -0.0050 -1.5562 -0.2613 -0.2100  0.7740  1.2549 -0.8484 -1.3120
#>  1.8029 -1.9986  0.2526 -0.0245 -1.0225  1.3654  0.4022 -1.0418  1.7148  1.3487
#> -0.2900  1.5543 -0.9827 -1.7007  0.0015  0.0026 -0.3780 -1.7286  0.3880  0.1710
#>  0.2789  0.9599  1.0091  0.6797  0.7321 -0.6337 -0.0583 -0.8344 -0.0844  0.4842
#>  1.2703  0.4395  1.4611 -2.6483  0.3733  1.2161 -0.2710  0.8885 -0.2597  0.4582
#>  1.0932 -2.2764 -1.1051  0.8095 -0.5097  0.4600  1.2841 -1.9238  0.5708 -0.7274
#>  1.3257  0.4114  1.0176 -0.3005  1.5741  0.9062 -0.3042 -1.7689  1.4688  0.1515
#> -0.2488 -0.2013 -1.2251  0.4920 -0.5195  1.3516  2.0782 -0.7077 -0.0902 -0.8179
#> -0.4217  1.3896  0.1071 -0.8734 -0.8015  1.4812 -1.0716 -1.4416  0.7633 -0.6735
#>  1.5116 -1.4046 -1.8321  0.7567  0.6818  0.1083 -0.5007 -0.3184  1.0954 -0.3198
#>  0.3164  1.4104 -0.6964 -0.9184 -0.7262 -1.0532 -0.6103 -1.4671 -0.4776  0.2094
#>  0.6300 -1.1566 -0.4814 -2.3693 -0.4130 -0.0877  1.0710  0.9140  0.2106  0.0011
#>  0.9653 -1.0773  0.2332  0.2301 -0.4845 -1.2533  1.0774  1.2599  0.7320  0.4795
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]