Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.6245 -0.8190  1.4157  0.0143 -0.4855  0.1312  0.1520 -0.3763 -1.4977  1.4714
#>  0.8322 -1.2783  1.3751 -1.1514 -0.8201  1.6119 -2.1686 -1.4568 -0.2942 -0.8755
#>  0.3107 -1.6895 -0.2646 -1.2348 -0.1702 -0.4347  1.4401  0.7699 -0.9657  0.5206
#>  0.3636  0.6527 -0.3112 -0.1807  1.0061 -1.7128  1.1611  1.0522  0.1356 -0.5512
#>  1.3212  0.2095 -1.0903 -1.2175  0.2315  0.4522 -1.6835 -1.4562 -0.8145 -0.2649
#>  0.5331 -1.2405  0.2643 -0.9219 -0.3361  1.4205 -0.1752 -0.3935  0.8514  0.9118
#>  0.9402 -1.8287 -0.7733 -1.2214 -0.7494  0.9483 -0.3280  0.7095  0.2434  0.3176
#>  0.0897 -0.1391  0.0146  0.5939  1.4047 -1.8136  0.0628  0.0067 -0.4128  0.5705
#>  1.2731 -0.6166  1.7613  0.3799  0.0444 -0.1039 -1.5314 -1.7600 -1.3865 -0.3071
#> -0.7645 -0.4507 -0.6982 -0.5591  1.3661 -0.8980  0.0345 -1.7751 -0.7780  1.1741
#> -2.5975 -1.0749  0.0848 -0.3496  0.6449  0.1843  0.4212 -0.7240 -0.6472  0.5478
#>  0.0041 -0.4157  1.2551  1.9164  0.1608 -0.0986 -0.1141  0.2221  0.4225  0.2236
#> -0.5301  1.4080  0.4545  0.6229 -0.5197 -0.1329  0.6776 -0.1522 -0.8993  0.7962
#>  1.3910  1.1999 -0.6089  1.8260  0.9686 -0.0275 -1.1934  0.8299 -0.0338  0.2062
#>  0.7500  2.3046 -0.8399 -0.1023 -0.1336  0.4833 -0.4302  0.6725  0.1098  1.4861
#> -0.1611 -1.7402  0.2226  1.0760  0.7282  0.7155 -1.6834  0.4749 -0.8850 -1.3906
#>  0.9628 -0.5927  0.6814 -0.4210  0.7595  0.8263  0.6526 -0.4564 -0.1726 -0.1129
#> -2.1483 -0.5485 -0.2050  0.2905 -0.3903  0.8728 -0.6394 -0.1392  1.0116  0.9983
#>  0.2937 -0.2375  1.0337  0.4492 -0.1229 -0.4811  1.5226 -0.4622 -0.5893 -1.6087
#>  1.6544  1.0998  0.6742  0.6604  0.3592  1.4829  0.7183 -0.4225 -1.5870 -1.0177
#>  2.4592 -0.7154  0.3127 -0.0083  0.0818 -1.0105  0.4509 -0.2402 -0.0933  0.3492
#> -0.5254  1.6954 -1.3911  0.8087  1.2232 -0.0398 -0.4757  0.2121  1.5689  0.0201
#>  1.3447  1.1679  1.4519  0.4641 -0.3575  0.3739 -0.3255 -1.1300  0.4067  1.0754
#>  0.9899  0.2161 -0.5706  0.1773 -0.5598  0.6889 -0.8079 -1.3503  0.2964  1.3443
#>  0.6029  0.0570  0.3346 -0.2759 -1.5263 -0.1848  0.3283  1.0615 -0.7335 -0.2592
#>  0.7644 -1.5009  1.9167 -1.8773 -0.2536 -0.8861 -0.5840 -0.0235 -0.3920 -1.3279
#>  0.7254 -0.6963 -0.7805  0.7712 -0.2119  2.4409  0.1283  0.0209 -0.0092 -2.5093
#> -0.4328  0.0874  2.2130  0.4574  0.4892 -1.0859  0.9047 -2.3225 -0.1638  1.2117
#>  0.2082  1.4420  1.6050  2.5158 -0.4529  1.1216  0.7773 -0.2171 -1.9183 -0.5606
#>  0.4454 -1.2510 -1.1095 -1.2018 -0.7031  2.3337  0.1321 -0.2299  0.7850 -0.1652
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]