Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.2759 -0.8523 -0.4391 -1.2363  0.0075 -1.2621  0.1737  0.8542 -0.5719 -0.3808
#>  0.7613  0.9028 -0.6075 -0.5654 -0.4051 -0.2007 -0.4649  1.6981  0.2451  2.8038
#>  1.1373 -1.9726  1.1252 -0.5979  0.8442  0.4295 -0.0898  0.1736  1.4542 -0.0516
#>  0.6926  1.1624 -0.5979  0.0183 -0.4747 -0.3411  1.0187  0.2809 -1.2799 -0.3723
#> -0.5844  0.0870 -0.5371 -0.6386 -0.3048  0.9001  0.1659 -1.3902 -0.0637  1.3400
#> -0.6414  0.0699  1.0686  0.4360 -1.1876  0.4948 -0.2422 -0.5290  0.5978  0.3215
#>  0.2550 -0.0517 -0.5026 -1.3010 -0.6436 -0.1340 -0.2519 -0.8465  0.7841 -0.2820
#>  0.6445  1.1559 -0.5948 -2.1692  1.1100 -1.0945  0.8385 -1.6010 -0.9807  0.3390
#> -0.5975  0.8840 -1.7726  0.8969  0.4778  0.5285  0.4279  0.6504  2.4500 -0.0462
#> -1.4533 -0.6604 -0.5455 -0.5988  0.7253  0.3425 -0.3876  1.3493 -0.0631 -1.0857
#> -0.9133 -1.0859 -0.4418 -0.3031 -0.0439  0.5192  0.3944 -1.0966 -1.1943 -1.0776
#>  1.0420 -0.0718 -1.3684 -1.0904  0.9721 -0.9916 -0.7545 -1.1415 -1.2295 -1.7400
#>  1.2799  0.1625  0.2365  1.2897  0.8107 -0.2935 -1.8188 -0.7707 -1.0404  0.1653
#> -0.3026 -1.2364  0.3185 -0.6973 -3.4586 -0.2498  0.0762  1.3970 -1.4826  0.1846
#> -0.8688  0.1490 -0.0141  0.4518 -1.0316 -2.3264 -2.1577 -0.4039 -0.3299 -1.5507
#> -2.3279  0.5822  1.3319 -0.0106 -0.2716  0.8320  0.7263  0.3997 -1.1748  1.1031
#>  1.8266  0.1644  0.7959 -1.3095  1.0546 -1.0853  1.2792 -0.5806  0.0875 -0.1881
#>  0.1074  0.3020 -0.1251  2.8150 -0.1496 -0.9565  0.6758 -0.6444  0.1806 -0.3159
#>  0.4307  0.4642  0.0338 -0.7102  0.1039  0.5235 -0.6956 -0.8356  0.3089  0.6753
#> -0.3085 -0.0536 -0.6198  0.7878  0.0601 -0.7668  0.0955 -0.7172 -1.7925  0.7246
#>  0.0457 -1.0736 -0.1108  0.5794 -0.3536 -0.5317  0.5628  0.2286 -1.4945  0.7771
#>  1.5771  0.0815  1.1345  1.1289 -1.2399  0.5777  0.4717  2.1672 -0.8149 -1.5109
#> -0.1324 -0.7494 -0.2990  0.3835  1.1024  0.9149 -0.9528 -0.6246  0.1078  0.8679
#> -0.7375  0.9850  1.8822 -1.7305 -0.9533 -0.4838 -0.5366  2.4077 -0.1297 -0.1768
#> -0.3789 -0.7361  0.1170  0.9928 -2.6739 -0.7179  1.6348  1.6835  1.4720  1.1073
#> -0.2543 -1.1334 -0.1477 -1.6088  0.3341 -0.4097  0.2883  0.6946  2.0801 -0.1747
#> -0.1967  1.2354 -0.3836  0.0317 -1.1089  0.9839 -0.4531 -1.5945 -1.1606 -1.5646
#> -0.6276  1.4209  0.2539 -1.4986  2.5493  0.1340 -0.0071  0.0731  0.5649  0.7044
#>  0.7290 -0.7754 -1.0652 -0.2899 -0.4232 -1.2521 -0.8118 -1.0367  0.2651  0.9868
#> -0.0588 -0.3315  0.5849 -2.0152 -0.4800  0.9806 -0.2563  0.4702  0.3338 -0.9354
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]