Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.2670 -2.4203 -0.3495 -1.0157  0.4542  1.1669 -0.7974 -0.1998  1.4366  0.5920
#>  0.0458  0.1138 -0.3388  0.6979  1.9001 -0.0920  0.0257  0.3075  0.7157 -0.1393
#>  0.3030  1.0364 -0.3448 -0.6398 -0.0135 -1.2373 -0.0744  0.0409 -0.3158  1.2279
#> -0.1536 -1.2160 -0.9250 -0.6840 -0.6317  0.6179  0.7305 -1.6056 -1.3159 -0.1836
#>  0.2731  2.4563  0.5035 -0.8574 -0.3192  1.0530  0.5919  0.1275  2.3479  0.3256
#>  0.4467 -1.0017  1.5469 -0.0339 -0.7140  0.4244 -0.7653 -2.8908  1.7167  0.0506
#> -0.3445  1.9997 -0.8283 -1.6467 -2.0765 -0.7904 -0.2495 -2.3882  0.8783 -0.4190
#>  1.5218 -1.0385 -0.5224  1.0800 -0.5652 -0.5356 -1.4899 -0.0743  1.1712  0.5023
#>  0.9875  1.0387 -0.9844 -0.8195 -1.0218  1.3591 -0.8253  0.2219 -0.9670 -0.3655
#>  0.0313 -0.5566  0.3656 -1.1927 -0.0146  0.4339 -0.3042  0.3287  2.2664 -0.2568
#> -0.0161  0.2408 -0.1133 -0.7077  0.2981 -0.6567 -2.7027  0.0327 -0.1378  1.2677
#> -0.5260  0.3808  0.1371  0.5462  0.0802 -1.1464  1.0169  0.0790 -0.1310 -0.3632
#> -0.8818  0.4766  2.1712  1.7547  1.4950  0.8595  0.3485 -0.3198  2.0683  1.4796
#>  0.7770  1.0971  0.7355  2.2728  1.5549 -1.2133 -1.0420 -0.5337  1.3366 -0.6936
#>  0.6623 -0.1682  0.5837 -1.9416 -0.5183 -0.4979  0.4329  0.7184  0.1303  0.5835
#> -1.1876  0.2919 -0.1455 -1.3536  0.6329 -0.5641  0.0464 -0.1631 -0.7572  1.1981
#> -0.4273  0.1015 -0.7950  1.8397  0.8920  1.0038  1.2018 -1.1805  1.6445 -2.0578
#>  0.3528 -0.4600  2.0719  0.7118 -0.3658  1.2602  1.1121  0.1429 -0.2903 -0.2133
#> -0.4832 -0.6420  0.3537 -2.1039 -1.9105  0.9928  0.9829  0.1323  0.7606 -1.6233
#>  0.9079  1.0434 -1.2854  1.6415  0.3935  0.3458  0.4855 -1.0381 -0.7139 -0.4941
#>  0.9538  0.9747 -0.3947  2.2265 -0.2203 -0.6869 -0.2110  0.5786 -1.0848  0.8089
#> -0.1883 -1.5878  1.0765 -0.4947 -1.7933 -1.2279 -1.4658 -0.3926  0.2412 -1.5334
#>  1.3340  1.9200 -1.3366  1.0979  0.8697  0.5531  0.4880 -0.1143  1.9801 -2.3437
#>  0.7913 -1.1198  1.1605  0.8077 -1.5695 -1.2313  0.3560  0.7079 -1.5469  1.2760
#> -0.2846 -0.8079  2.0678  0.5304 -1.4948  0.2175  1.5201  0.1852  1.5877  1.4293
#>  0.5275 -0.6469 -0.7280 -0.3487  0.1279 -0.6151  1.1548  0.9399 -0.4661 -0.2161
#>  0.0289 -0.4015  0.2864  0.8529  2.2055 -0.0686  0.9189  0.2377 -0.4433 -0.9943
#> -0.4495  0.9601  1.2620  0.7436 -0.6272 -0.4843  0.1146  0.1029  2.5025 -0.9680
#>  1.4580  1.1372  0.0904 -0.9478 -0.6160  1.4418 -0.0784  0.3617  0.8510  1.1805
#> -0.7859  0.1409  0.7105 -1.8944  0.7672  0.2563  1.7106  0.1080  1.4160 -1.2078
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]