Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.2690  1.1280  0.9372 -0.4203  0.0873  0.4052  0.4117  1.3813  0.4216 -1.0136
#>  0.9325 -1.3453  0.9559  1.7954  0.0899  0.0712  0.1747 -0.4314 -1.4221 -0.9310
#>  0.4803 -1.0696  0.5350 -0.1910  0.7358 -0.8103 -0.1351  1.1713  0.4417 -1.2849
#> -0.8889  0.1128 -0.5005 -0.1454 -0.3602  0.7356 -1.2586 -0.3340 -1.1571 -1.4277
#>  0.3315 -0.4910 -0.1480  1.5240 -0.0109 -0.7486  1.0323 -0.4272  0.6256 -1.4957
#> -0.6662  0.1483  0.8785 -0.3227  1.0425  0.3952  0.6237  0.7391 -1.1671 -3.4670
#>  0.2432  0.8028 -2.5756 -1.0251  0.0525  0.8127 -1.9401  0.1420 -0.1960  0.0696
#> -0.2430 -1.6470 -1.4150 -0.9568  0.3692  0.2499  0.1998 -0.8819 -1.5393 -0.3328
#>  0.5328 -0.2190 -0.4378  1.2280  1.1898 -0.8890 -0.1282  0.9241  0.9454 -0.2404
#>  0.3303  1.2129 -0.2446 -0.0324 -0.7455 -0.3890 -1.1760 -1.1805  0.3707  0.2394
#> -0.5921  0.4983  1.3128  0.4063  1.5903  0.7906 -0.5948 -0.8373 -0.5998 -1.3513
#> -1.4986 -0.6995 -0.3077 -1.0480  2.5864 -1.4864  1.0304 -1.4305 -0.5960  1.0424
#> -0.9420 -0.4043 -0.5590 -1.2799  1.9643 -0.6884  0.4628  0.0325  0.6009  0.8695
#> -0.8590 -0.3046  0.5777  0.9412  1.1758 -0.9984  0.1105 -2.4552 -2.4998  0.9501
#>  1.6258 -0.6202 -1.0380  0.3579 -0.1052 -0.7904 -0.7615 -1.0874 -1.0962 -0.4756
#> -0.1636 -0.2274 -0.5665 -1.3459 -0.1370 -1.4739  0.6973 -1.0108 -1.8132 -0.5133
#> -0.4516 -0.6729  0.8097 -0.6464 -0.8400  0.2059 -0.7398  1.3130 -0.1542  0.1568
#>  1.6202  0.8268  1.3221 -0.2814  0.7262  0.9758 -1.1984 -0.2714 -1.9382  0.4518
#>  1.5186  1.6847 -1.3189  1.0098 -0.5440 -1.3555 -1.2027  1.6921  1.1021 -0.9086
#>  0.5377  0.3504 -1.7649  0.9269  0.4620 -0.1836  1.4243 -2.8279 -0.6140  0.5761
#> -0.9359 -0.1190 -0.6524 -1.2038  1.0315 -1.7886 -1.1864 -0.7946  0.5358  1.1128
#> -1.4269  1.3659 -0.7958  0.4646 -0.4640  0.3053  2.6626 -1.3312 -0.4056  1.3496
#>  0.3805  0.2416 -1.0867  0.6434 -0.5970 -0.0584  0.8897  1.0362 -0.5001  0.4825
#>  0.4099  0.7695 -1.3595  1.4745 -0.6087 -1.2520 -2.0116 -1.3136 -1.2764  1.5341
#> -0.0523  0.8882 -0.8163 -0.7920  0.3148 -0.1117 -0.2588 -0.6825  0.1412  0.5551
#>  0.3318 -0.8655 -0.7052  0.0241 -0.2452 -1.2199 -0.5639 -0.4064 -0.7786 -0.6633
#>  0.7771 -2.1117  0.3364 -0.3983 -0.1659 -1.8291 -1.2770 -2.1136  2.0500 -1.6734
#>  0.2594 -0.2899 -1.8881  0.1440 -0.3125  0.1660  0.0432  2.3352 -0.6641 -1.2116
#> -0.9496 -0.3616 -0.4671  0.2819  1.1087 -1.2783 -1.4436 -0.9021  0.4040 -1.7140
#> -0.0724  0.5148  1.0298 -2.4110  0.7009 -0.5073 -1.3397  1.6912  0.5331 -1.4461
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]