Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.3138 -0.8611 -0.4446 -0.5490  1.0109  1.1561  1.1475  0.4488  0.6495  0.8498
#>  1.1211 -0.5694 -0.8023 -1.2484 -0.7287  1.3537 -0.5021  0.6010 -0.4883 -0.6482
#>  0.9905 -2.1082 -0.0251 -0.4411 -0.2412 -1.2712  0.0769 -0.7792 -0.6760  1.8735
#>  1.6725  0.2535  0.9489  2.2129  0.9543  1.9222  0.5306 -0.9067  0.3690 -0.4624
#> -0.5583 -2.0142  0.8820 -0.7926  0.8771 -1.1459  2.0283  0.3460 -0.6768 -0.0119
#> -0.6401 -0.1432  0.0684  0.0519  2.2383 -0.2797  0.4845  0.6395 -0.8397 -0.7347
#> -0.6527 -0.1114 -0.0638 -1.5564  1.2693 -0.5132  1.0853  0.6693 -1.0114  0.3000
#> -0.3996  1.7450 -0.2072 -0.6019 -1.9791 -1.1230 -1.1329  1.7149 -0.6946  0.3007
#> -1.6643 -1.3994  1.0635 -0.3972  0.7962 -1.3295 -0.2794  0.2068 -0.1009  0.7874
#> -0.5121  0.0983 -0.5896  0.7636  1.2964  0.2895  1.1722 -0.6817  0.1888 -2.5483
#> -0.1977 -0.4903 -0.3203  0.3551  1.2609  0.1602 -0.7972 -0.4104 -2.0912 -0.0299
#>  1.0512  0.5748 -2.3604 -0.9936 -0.3276 -0.0122  0.2952 -1.0132  0.8601  2.6199
#> -0.1176 -1.7705  0.0118 -0.1168 -0.3461 -0.0462 -1.2710  0.4167  0.7105 -0.5731
#> -0.8555  0.4739 -1.1143  0.3451  0.5458 -0.0653  0.1475 -0.5787  0.3016  1.0060
#> -1.3316  0.1540  0.0823 -1.9769 -0.1247 -0.4997  0.3527  1.8233 -0.1186  0.5522
#>  1.4993  1.2426 -0.7349 -0.5448 -1.0628  0.2141 -0.5620 -1.2196  1.2271  1.0073
#>  0.1695 -0.1115  0.7127  0.3665  1.0320 -0.0099 -0.3279 -0.5304  0.7258  0.5766
#> -0.8967 -2.0989 -0.2017 -0.1703 -0.5276 -0.5828 -1.6947 -0.1102 -0.1229 -1.0758
#>  1.2755 -0.0066  0.5129 -0.7379 -0.9161  0.3911 -0.9651 -0.5953 -1.1507 -0.9622
#>  0.7008  0.4327 -2.1556  0.1010 -0.7420  0.7267  0.1249 -0.6432 -1.2108  2.3390
#> -1.0243  0.5903  0.0109 -0.1741 -0.4324 -1.1883 -0.8767 -1.4313  0.7849  0.2774
#>  0.0149 -0.4148  1.9616  0.2927 -0.1670  0.0620 -1.6566 -1.3382  1.5290 -0.9194
#>  2.1055 -1.3546 -0.8530  0.5457  2.7790  0.9264  0.3293 -0.0126 -1.1341 -0.7563
#>  1.4824 -0.0754 -0.3891 -0.7064 -0.6220 -0.7225 -1.5662 -1.0223 -1.4600 -0.2370
#>  0.6369 -0.6457 -0.7759 -0.1725  0.9587 -1.5994  1.2848  1.0212  0.4079 -1.4350
#>  0.1364 -0.5716 -0.3221 -1.1412  1.5517 -0.3213 -0.0857 -0.3433  0.4506  0.0016
#>  0.0182 -0.5257  0.1121  1.1386  0.2836 -0.9591  0.1188 -1.4842  1.6995 -0.1127
#>  0.3963  0.8050 -0.8508 -0.8576 -1.2801 -0.9357  1.0174  0.7109  0.0867 -0.0898
#> -0.3821  0.2749 -0.2067  0.7027  1.0554  0.7722 -1.0461 -0.4867 -0.1568 -0.7082
#> -0.6330  0.8016  0.2440  0.3676  1.0784 -1.4064  0.2184  0.1433 -0.5694 -0.3184
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]