Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4442  0.5598 -1.6156  1.0801  1.5944 -2.4856 -1.1014  0.5966 -2.5073 -0.1099
#>  0.4261 -0.3354  0.1689  0.8440  0.9565  1.2103  0.9024 -0.7263  0.7619  1.6570
#>  1.5267 -1.3212  0.5115 -2.0473 -1.4922  1.0575 -1.2116  0.1369  0.1236 -0.5169
#> -0.1319  0.0853  0.5291 -0.4967  0.6120 -0.1695 -0.2821 -0.9244 -1.2769 -0.8735
#> -1.6178 -0.7428 -0.3841 -1.1757 -1.0454  1.7176 -1.0539  0.3697  1.2667  0.8133
#>  0.1657  1.0800  0.8516  0.7288  1.1392 -0.6573 -0.2139  0.3435 -0.2722 -0.9175
#>  0.3248  0.6373  0.2629 -1.1282  1.5019  1.0652  0.4313  1.0741 -0.5733  0.0833
#> -0.5988 -0.8280 -1.6267 -0.4810 -0.3993  0.7342  2.0969 -0.6718  0.2776  0.7419
#> -0.8116  0.4262 -0.1465 -0.5310 -1.6563 -0.2683  0.6270  0.6725 -0.0117 -1.4937
#> -0.1232 -0.7565 -0.2914 -0.6949  0.6559 -0.1339  0.6372  0.6918 -0.0704  0.2217
#>  0.1609 -0.5689 -0.1542  0.2805  1.3068  1.0518 -0.0973  0.3120  0.6408 -0.8671
#> -0.4712 -0.9209  1.7343  1.1586  0.2941  0.5936  0.5828 -0.2025 -0.4864 -0.2739
#> -1.4013 -1.8266  0.6959 -0.3775  1.6589 -1.0765 -0.9526 -1.3362 -1.3533 -0.8948
#>  0.6042 -0.4399 -1.4612  0.3398  1.3375  1.1032  1.9413 -1.0198 -0.0797 -1.1650
#>  0.9095 -0.8647  2.7712  0.9975  0.4208 -0.5663 -2.2044  2.3612 -0.3409 -1.2816
#>  0.8598  0.6798 -0.6627 -0.5573  1.8040 -0.2105 -0.4373  0.5921  0.9355  0.0920
#> -1.6371  0.3031 -1.6970 -0.5414  1.8905 -0.1673 -0.8990  0.9148 -1.1247 -1.3663
#>  0.4759  0.7160 -0.2455  1.5496  0.8228 -1.3361 -1.0104 -0.0378 -0.9149  0.1983
#>  0.2292  0.3399 -2.5734  1.5658 -0.5052 -0.0766 -0.7858 -0.1965 -0.2207  0.3935
#>  0.8521  2.2912 -0.8789 -1.3700 -0.6968 -1.0259 -0.9962  1.6068 -1.2262 -0.4789
#>  0.9862  1.2074  0.5145 -0.7367  1.8433  0.6130  2.2086 -0.8758  0.2375 -0.7679
#> -2.0040  1.3510 -0.5073 -0.2409 -1.4248 -0.0737 -0.7916 -0.4155 -0.0337  0.8926
#>  0.6815  0.1904  0.8337 -0.9652  1.7863 -0.2759 -1.8037  0.2193  0.5788 -0.6576
#>  0.4212 -0.4255  0.6080  0.1356 -0.0485 -1.2374  1.1070  0.6060  0.2132  1.3261
#>  0.5166  0.0186  0.2662 -0.2361 -0.4313  0.0634  0.3090  1.6845  0.4470 -0.0680
#> -0.1342 -0.2884 -0.0495  1.3609 -0.4599 -0.2347  1.1448  0.2784 -0.1310 -1.1163
#>  0.0678  1.2702 -0.8139  1.0605 -1.0842 -0.3317  0.6538  0.8407 -0.4327  2.3734
#>  0.6269  1.3184  1.1240 -0.5458 -0.1290  0.3598 -1.6860  0.8951  0.4242  1.8250
#> -0.2993 -1.6450 -0.6424 -0.5554 -0.3011 -1.6037  0.2557  1.2804 -0.3786  1.4816
#>  1.7621  0.0092 -0.7663 -1.6827 -1.0651  0.6314  0.2747 -1.3768  1.0451  0.5036
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]