Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4250 -0.2232  1.2960  1.7756 -1.6423 -0.9362 -0.4446  0.3400 -0.3193  0.2449
#>  1.0053 -0.4777 -0.5802  0.0650  0.4738  1.3754 -0.1927 -0.7199  1.6885  2.3652
#> -1.0696  1.8033  0.0830 -0.7397  0.4662 -0.4605  0.2509  0.3753  0.0465 -0.4119
#>  1.2087  0.3957 -1.8711 -0.8905  0.3510  1.6983 -0.5601  1.1172 -0.2835 -0.9423
#>  0.8823  1.0405  0.3815 -1.6763  1.5335  1.4310 -0.2703 -0.3092  0.3801  0.3449
#> -0.0298  0.8337 -1.5782 -0.8332  0.8572  1.3617  0.3539  0.0148 -0.1587 -0.1884
#>  1.4208  0.2059 -0.3856  0.0737  0.2866 -0.2439 -0.2358 -0.7357 -0.9891 -0.6607
#>  2.3359  0.2415  0.9637 -0.1562 -0.1343 -1.8927  0.9750  0.2841  0.5336 -0.9225
#> -1.5637 -0.8068 -1.2038 -0.0402  0.6359  1.4984  0.3242  0.1863  0.0623 -0.5787
#>  1.0390  0.9608 -0.7734 -1.0311 -0.3641 -1.1466  1.0800 -0.0500  1.8431 -1.7728
#>  0.3127  1.5569  0.2050 -0.8157  0.1371  0.5650  0.6797 -0.2100  0.8504  0.2863
#>  0.1299  0.4832 -1.3099  0.0206 -1.4550 -0.2872 -0.2414  0.7534 -0.3200 -3.0479
#> -0.7229 -0.0028  1.0825  1.8853 -1.1355  0.9795  1.0620  0.7515  0.0608  0.2805
#>  0.6223 -0.0856  0.6364 -1.3439  1.0156  0.3500 -0.8009  0.5914 -0.1384  0.4511
#>  0.3935 -0.1673  1.0778  0.2059  0.3533  1.1779  0.0866  0.5286 -2.9507 -0.7650
#>  1.8775 -0.9880 -1.0030  0.9447 -0.3543 -1.4358 -0.4242  0.4002  0.7813 -0.3779
#> -0.4357 -0.2456 -0.5804 -1.6668 -0.1601  0.0576  0.4516 -0.2642  0.3330 -0.6564
#> -1.8673 -0.3349  0.0240 -0.9492  1.5236 -0.6889 -0.2804 -0.6716  0.6339  0.6518
#>  0.6305 -0.7113 -0.6518  1.5874 -1.7618 -0.1074 -0.4563 -0.9321  0.8192  0.8761
#> -0.5389 -1.4163  0.6260 -0.1096  1.4365 -0.8303 -0.1166 -0.4863 -1.6202 -0.6453
#> -0.4017  0.4011 -0.5204 -0.1747 -0.5427  1.9058  0.5464 -0.1802 -0.1019 -0.9003
#> -0.4605 -0.4352 -0.9525  0.2643 -0.3114  1.7017  0.1536 -0.4565 -1.2030 -1.2765
#>  0.7422 -0.4704  0.1782 -1.1222 -0.3406  1.0622  1.0285  1.2788 -1.7862 -0.1513
#>  0.6411  2.1271  1.5602  0.8308 -0.9556  2.5146 -0.5295  1.1288 -0.6051 -0.8543
#> -2.2444 -2.1560 -1.1865 -0.0126  0.9122 -0.8056 -1.5664  0.0827  1.4128 -1.4103
#> -2.0467 -0.0964  0.2102 -0.5717 -0.1520 -0.3014 -1.6804  1.7347 -1.5158 -0.9355
#> -1.2573  0.5778 -0.4617 -0.5478  0.4734  1.6987  1.3201  1.0416 -1.7623  0.0798
#>  0.7295  0.6363 -0.4067 -0.3609  0.3469 -1.4800  0.4310 -0.2303 -0.2128 -0.6119
#> -1.4355  1.6462 -0.5343  0.9907  0.0478 -0.0662  0.1174  1.5698  0.6730  2.1615
#> -1.5506  1.2938  0.7116 -0.3555  1.6802 -0.0309 -0.3959 -1.9177 -0.0067  2.0742
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]