Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.3976  1.4102 -0.1612  0.7649  0.5638  1.6266 -0.1730 -0.0268  0.5017  0.0071
#>  1.4194 -0.1480 -0.0116  0.3978  0.4215 -0.8790 -0.8124  1.2814 -1.0666  0.7622
#>  0.7008  0.1857 -0.4897  0.3514  1.6323  1.6976 -1.2723 -0.8498  0.4065  0.5432
#>  3.0036 -1.4415  0.1982 -1.5013 -0.9881 -0.8389  1.4030 -0.5013  0.0872  2.3395
#>  0.3796 -1.2730 -0.6959 -2.1502 -0.9181  0.9935  1.3448 -0.6070  0.1379 -0.9692
#>  1.2739  0.5741  2.1181  1.8979  0.7053  0.5165 -0.0388  0.0797  1.1330  1.1257
#>  0.1509  2.0685 -0.0873  0.6388 -1.2580  0.2487 -1.1275 -0.2883 -0.0319 -0.1253
#>  0.8429  0.8222  1.9009  1.9415  0.2682  1.2596  1.2404 -0.5446  2.6397  0.5771
#>  1.6352 -0.7642  1.9217 -2.8625 -0.8305 -0.0862 -0.8351  1.0810  0.1634 -0.9570
#>  0.0310 -1.5271 -1.0466 -0.4795  0.2388 -0.0815  0.2925  0.0537  0.4615 -0.9746
#> -0.6257 -3.2889 -0.1486  1.4016  0.4153 -0.2165  1.4627  1.3917  0.5899 -0.3035
#> -0.0177  0.3846  0.8480  0.0315  0.1583  1.1965 -0.2792 -0.1176  0.6512  1.3540
#>  1.1746  0.5607  0.3341 -1.6655 -0.6736  0.2536  0.2456  0.6536  0.7038  0.4018
#>  0.5948 -0.0341  1.1356 -0.7041 -1.2234 -0.5339  0.1451  0.3979  0.6892  0.8465
#>  0.9016  0.1561  0.0625 -0.2514 -0.1027 -0.3792 -1.2673 -1.0544  0.0726 -0.4510
#> -1.2260  0.4953  1.6646  0.9323 -2.2194 -0.6554 -0.0639 -0.4152  0.3523 -0.5326
#> -0.9809 -2.4771 -0.9164 -0.5790 -1.7819  0.6417  0.5549 -0.9099 -0.7779  2.8152
#> -0.9554 -1.0458  2.0056 -1.1786  1.1714  0.4156  0.4425 -3.0495  0.7325 -1.0165
#>  0.2082 -1.0691 -0.7507 -1.9541  1.3131  1.4621  0.4408  1.6208  1.2499  0.5086
#> -0.7359  1.2203 -0.9366  0.8946  0.8530  2.1446 -1.2995  1.1173  0.2747  0.2998
#> -0.3094  2.2035  0.0654  0.7115  0.6325  1.5367 -0.7421  0.9115  0.5633 -0.8859
#>  0.5540 -1.3216 -0.3132  0.2844  0.6021  0.6180 -1.6759 -0.7701  2.2269 -0.2454
#> -1.4535  0.1063 -0.9981  1.4948 -0.3283 -0.2985  1.0425  1.0207  1.1077 -0.2802
#>  1.3363 -1.5279  0.2950  1.2896  1.6622  1.1163 -0.0310  0.4462 -1.3409  0.8673
#> -0.6279 -0.4254 -0.2153  0.9036 -0.5665  0.1212 -0.0982  0.8168  0.7073  0.8298
#> -2.1345  0.6199  0.4003 -0.1820 -1.2050  0.4819 -0.2807 -1.3277  1.0401 -0.2565
#> -0.8271  0.6493 -0.6739  1.7806  0.0078  0.2453  1.0074 -0.5097  0.0999  0.6490
#>  0.0500  0.4980  0.7713 -1.1428  0.3989  1.3204 -2.1891  0.7766  1.5256 -0.2772
#> -0.0834 -0.5852  1.9838  0.7738  1.3359 -0.1341 -0.7053 -0.5893  0.1170 -0.1613
#>  1.1466  0.5575 -0.0954  0.4604 -0.4136 -0.3532 -0.2871 -0.8054  1.6762  1.5391
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]