Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.6571  0.2224 -0.4773 -0.9379 -0.0261  0.9694 -0.4650  0.8224 -0.4611  0.4922
#>  0.2174 -0.6422  0.8584  0.6459 -0.6299  0.1085 -1.5910 -0.1507  1.5367 -0.8113
#> -0.1491  0.9655  0.2983 -0.4673  2.1705 -0.4620  0.1435 -1.0409  1.0935 -0.4218
#> -2.3423 -0.3543  0.8933  0.0917  1.7262 -1.6373 -0.7797  2.9204 -0.2081  1.7133
#>  0.0261  2.7617  0.8750  0.8941 -1.7339  0.5237  0.6518 -1.0848  0.6049 -0.7995
#> -1.3611  0.7824  0.5632 -0.9699 -0.8093  2.2999  0.1659  0.3041 -1.2897  0.0665
#>  1.6138  1.1423  0.0282 -0.8771 -0.9710 -0.0982 -1.4529 -0.5897  1.2027 -1.1056
#>  0.7225  0.3934  1.2461 -0.3570 -1.4605 -1.3378  0.3794 -0.3731  1.3891  0.6563
#> -0.7199 -0.4032  0.0833  1.1864 -0.2428  1.0876 -1.0572  0.6406  0.4197 -0.1609
#>  0.4696 -0.0659  0.1295  0.6661 -0.6669  0.2497  1.1722 -1.4193  1.8854  0.3679
#> -0.3591 -1.3209 -1.2409 -0.8989 -0.4355 -1.8653  0.5261  0.5451  0.9231 -0.1412
#>  0.7508  2.2535  0.0280 -0.6753  1.3405 -0.0287  1.2038 -0.0949 -0.1743 -0.9666
#> -0.9798  0.2576  0.8546  0.6891  0.0106  0.2847  0.4512 -1.1922 -0.4845  0.5365
#> -1.1307 -0.7625  1.5413  1.3094 -1.2690 -0.7578 -0.6237 -1.5197 -0.6827 -0.0090
#>  1.1445 -1.6202  1.6156  3.6099 -0.1821 -0.5620 -1.2963  0.4956  0.6879 -2.3588
#>  0.3693 -1.6377  0.1819  0.8796  0.2188  0.6769 -1.6577 -0.0112 -0.1871 -1.3408
#> -0.2271  1.6144  0.0861  1.9757  0.7142 -0.8741  0.5272 -0.6242 -0.4481  0.9671
#>  2.2840 -0.6375 -1.1349 -1.4904  0.1259 -1.3974  1.9710 -1.7029 -0.3481 -2.4434
#> -0.5392  0.7739  0.3489 -0.1937  0.0353  1.3455 -1.3567 -0.0467 -0.3981 -0.1427
#>  0.9104 -0.2840 -0.2618 -1.1080  1.4202  0.0211  0.0942  1.0768 -0.2419 -3.0529
#> -2.0401  0.5971  1.1173  1.0888  0.0758 -1.7036 -0.8627 -0.9335 -2.0346 -0.0880
#> -0.1252 -0.0147 -0.7374  0.7822 -0.5338 -0.5072 -0.0857  1.3305 -1.0437 -1.0978
#> -0.6000 -1.4271  1.9856 -0.1974 -0.8187  0.4682 -1.5175 -0.7640 -0.9971  1.4452
#> -0.4266 -1.1445  0.2219 -1.2331  1.0194 -0.0846 -0.3067  0.1594 -0.1166  0.1412
#>  1.3185  1.2201 -0.9208  1.0082  0.0517  0.4011 -0.3084  1.1075  0.1750 -1.0289
#>  0.5360  1.0408 -1.7380 -2.1046  0.9841 -0.0538  1.3321 -0.8003 -0.2867 -0.5408
#> -1.9512 -1.2466  0.4413  1.9485 -0.7929  0.5197  0.5425  0.6533  0.0979  0.4650
#>  1.5729 -1.1579 -0.3600  1.6203 -0.1021 -0.6264  1.5127  0.0864  0.0178  1.2972
#> -0.4037  0.3449 -0.1753  1.0235  0.2199  0.3373 -0.6978 -0.4742 -1.0005 -1.6060
#> -0.6623 -0.5052 -0.3875 -1.1880 -0.0096 -0.8448  2.0968 -0.2651  0.4241 -0.3058
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]