Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.3607  1.0990  2.3379 -1.1255  0.3079  1.3541  0.2643 -1.5434  1.0389  1.2821
#> -0.2826 -0.0538 -0.9729  0.4665  1.0320 -0.0593  1.3386 -1.1835 -1.6518 -0.9636
#>  0.7414  1.2645 -0.2784 -1.0694  0.7389  1.0459 -1.6370  0.6758 -0.5324 -0.0914
#> -1.6004  2.4927 -1.0092 -0.4040  0.7607  2.1137 -1.1725 -1.0582  0.4469 -0.1498
#> -0.6104  1.0233 -0.7256 -1.6990  0.9469  0.0512  0.9017  0.7913 -0.9768  1.4493
#>  0.7061 -1.0374  0.5466  1.8066  0.5542  1.2846 -0.9211  0.0484  1.5259  0.2726
#>  0.2341 -0.3996 -0.1335 -2.1585 -0.9433  0.4888  0.1860  0.3588 -0.7793 -0.2549
#>  0.9084  0.2636 -1.4510 -1.2449  0.7806 -0.0783  1.4700  0.7658  0.6775 -1.9940
#> -1.8855  2.1070  0.6444  0.1339 -1.3176 -0.5997  0.0648 -0.6032 -1.7431 -0.5366
#>  2.6643 -0.8656 -0.8306 -0.1812 -0.2974  0.4919  1.1295  0.6218 -0.4152 -0.6285
#> -0.9824 -1.6873 -0.6923  0.1105 -0.7974  0.9480  0.6564  0.2534 -1.2062  0.4254
#>  0.5078 -0.1750 -0.4173 -1.3898  0.8738  0.5606  0.5158  1.6087  0.9369 -1.1554
#>  2.0137 -2.6797  0.0122  1.1084  1.1134  1.0838 -0.6454  0.2782  1.0680 -1.0361
#>  0.1338 -1.5241  0.4264 -0.5999  0.9068 -1.3583  0.1442 -1.6829 -0.9397 -0.4897
#> -0.3954  0.1794  0.5829  0.0601  0.3521  0.0543 -1.2073  0.6801  0.3744 -0.5603
#>  1.9458 -0.9679  1.4183  0.0170  0.8852 -0.6699 -0.9052  0.9330  0.9293 -0.3607
#>  0.6876  0.2148  0.2113  0.7709 -0.6398 -0.6815  1.0777 -1.6991 -0.0127  0.6725
#> -0.3845  1.2874  1.7839  1.4397 -1.9468  1.0309 -0.2258  1.0239  1.2660 -2.2446
#> -1.2699  1.3756  0.7907  0.1188 -0.0269 -0.0812 -1.1396  0.7948  0.7120  0.3618
#> -2.1048  1.3926  0.4670 -0.3526 -0.0738 -1.0813  0.6002  1.7768 -2.4190 -1.3123
#> -1.0650  0.7594  0.6346  0.2392  0.1390  0.2440 -0.5244 -2.2370  0.3237  0.6650
#>  0.1037  0.6791 -0.2084  0.5807 -0.3763 -1.3375 -0.1968 -0.1434  0.5917 -2.0201
#>  1.5414  0.5828  0.5415  2.0649 -0.8655 -0.5345 -1.2156  0.3361  1.1245 -0.2776
#>  0.4254 -0.1977  0.5057  0.5778  1.7307 -0.7886  0.9186  1.4858 -0.1336 -1.3551
#>  0.0148 -0.9627 -0.1713 -3.6112 -0.4611  0.1148 -0.8404 -0.9799  0.4769  0.4456
#>  1.2153 -0.9135  0.9805 -0.0495  0.0784  0.4613  0.0606  0.5908 -0.0831  0.8864
#> -0.2163  3.0951 -0.6625 -0.8678 -0.0997 -0.3754  0.4516  2.5135 -2.1790 -0.2823
#>  2.5510 -0.9258 -0.2065  2.3025 -0.9761 -0.7360 -0.4868 -0.4429  0.4924  0.8370
#>  0.2747 -0.5033 -1.3399 -1.2163  2.6118  0.1069  0.8143  0.1894 -1.1606  1.8968
#>  2.2279  0.4697  0.1978 -0.2785  0.4038 -0.4434 -0.3244  0.8959 -1.3723 -1.2315
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]