Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.2000  0.7704 -0.3568 -0.5507 -1.4193 -0.3826 -0.3326  2.0335 -0.5317 -1.3395
#> -0.9310  0.3222 -0.3631 -1.3056 -0.8831 -0.1268  0.7918 -1.7166 -1.5391  0.0951
#> -0.7370  0.7949 -0.9402  0.7291  0.2600 -1.5195  1.4024 -0.8154 -1.6492 -0.4859
#>  0.7607 -0.5591 -1.3018  0.7704 -1.2026 -1.0272  0.1373 -0.9727 -2.0273  1.1828
#>  2.3147  0.2781 -0.5193  0.5602 -0.0542  0.4099 -0.1738  1.9531 -1.5029  0.9364
#>  0.2085 -0.8971 -0.9552  0.4372  0.6141 -1.1004  0.2221 -0.9672  0.6771  3.1025
#> -1.0878  0.5914 -0.5479 -1.9702 -0.2170  0.2107 -0.1795  0.0819 -0.2653  0.9427
#> -0.8750  0.2682  1.2989  0.1000 -0.9208 -0.6927  0.7887  0.1091 -2.0229 -0.8954
#>  0.1804 -0.0588  2.0295  0.3679  1.1628 -1.1475 -0.2251 -0.3888  0.3545 -0.8193
#> -0.1255 -1.8586  0.4109  0.9278 -1.1892 -0.4175 -1.4385  1.3519  0.2406 -0.3539
#>  0.7301  1.4874  0.5906 -0.1368 -0.4006 -0.2176 -1.0197 -2.6179  0.5397  0.3346
#> -1.6859  0.4825 -0.0182 -0.0967 -1.1116 -1.0240 -1.3152  0.7182  0.0393 -0.5831
#>  0.5543 -0.3184 -1.1745 -0.4625  0.1712  1.2906  0.5392 -0.7009  0.2113  0.6950
#> -1.3727  1.9526  0.0719 -1.8284  0.2061 -0.0263  0.0881 -0.1049  0.7131 -0.4704
#> -0.1352 -0.8363 -0.1987 -0.3452 -0.8955  0.4957  0.4515 -0.4342 -0.2373 -0.9015
#> -1.6942  0.3149 -0.6018  0.4863 -1.0642  1.7315 -0.2356  0.1427 -0.1572  0.4852
#> -1.4023 -0.6961 -0.5014 -0.7192 -0.5580  1.2032  0.3046  0.1389 -1.7287  0.7858
#>  0.6735  1.2783 -0.7162  1.1549 -0.1059  0.0363 -1.1343 -0.5115 -0.4988 -0.9344
#> -0.0069  0.6285  0.2213  0.7669  0.2030 -0.7022 -0.5677  0.4306 -0.0459  1.3606
#>  0.5373 -1.0789 -1.2484 -1.0774  1.1975  0.7493 -0.1142 -1.0336 -0.1629 -1.0193
#>  0.6287  0.4843  0.1196 -0.9777  0.8511 -1.1792 -0.5491 -0.6117  0.1468 -0.5619
#> -1.9745  0.2010 -0.4852  0.5633  0.5862 -0.1178  2.0722  0.2911 -0.2454  0.2077
#> -1.0937 -0.9987 -2.2484 -0.4275  1.2687 -1.0582  0.4915 -1.0082  0.1712 -0.5679
#> -1.1530 -0.2766  1.8968  1.2185  1.0450  0.2712 -2.3507  0.0154 -1.2662  0.1239
#>  1.3383 -0.8352 -0.0480  1.4574 -0.5519 -0.0289  0.7627  0.3051  0.0991 -0.3306
#>  1.2381 -1.3060  0.5424  0.3801 -0.0120 -0.2464  2.1142 -0.7269  1.1574  0.1200
#>  1.5047  0.3050 -1.8850 -1.0952 -0.1819 -0.6869  0.3749 -1.3044  1.0577  0.0482
#> -1.6080 -0.5829 -1.9385  0.2962  1.1991  1.8674  0.5164 -1.7146  1.8312  1.2178
#> -0.6023  0.7802 -2.5175 -0.5644 -0.1250 -1.2860  0.6245  0.4901  0.3913 -1.4470
#> -1.6349  0.5274 -1.0229  1.1077 -0.2156 -0.1541 -1.4945  0.9055 -0.2271  0.8295
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]