Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.2840  0.1797 -0.3339  0.0146  0.6145  0.6213  0.9106 -0.3376  0.1509 -0.7408
#>  0.6681  0.7294 -1.0805 -0.2449  0.7853  0.5467  1.5132 -0.5336 -1.1667 -0.6973
#>  0.3255 -0.5428 -0.4380  0.5104 -0.6511  1.7597 -0.2137  0.5435  0.3250  0.5606
#>  0.3602  0.7970 -2.7586  0.5749  0.2002 -0.7986 -0.4171 -0.2950 -0.0821 -0.9252
#>  1.7565  0.6962  0.7607  0.6603  0.3057  1.2664 -0.5413 -1.6286 -0.3151 -1.9112
#>  0.2634  1.5329  1.5294  0.3735 -0.0418  0.5037 -0.6230  0.4858  0.1772 -0.3002
#>  1.5456 -0.1444  1.3689  0.9863  0.3514  0.2611  0.0106  0.3838 -0.2301 -0.9887
#>  1.0602 -1.2040 -0.7563 -0.1157  1.2294 -0.4598  1.2365  1.1124  1.4183 -0.5609
#> -1.4800  0.8403 -0.5031  0.3209 -0.8140  0.3130  0.3919  0.6575 -1.2270 -1.4270
#>  0.1586 -1.1026  0.7231 -1.0585  0.6514  0.4498  0.0719 -0.1463  0.2662  0.9454
#> -1.1749 -0.7341 -0.1954 -0.0735 -0.8869  0.9992  0.5724  0.7904 -0.2105  1.3841
#>  0.4082 -0.3864  0.0358 -0.1917  1.5883  1.4514  1.1247  0.7979 -0.5325 -0.0563
#> -1.0745  2.0096 -0.3497  0.0700  2.9537 -0.8535  0.0553 -0.3942  0.0040  1.2763
#>  0.6271  0.4102 -1.4651  1.2230 -0.3347  0.0446 -1.5544  0.5172  0.9441 -1.8373
#> -0.6829  0.3948  1.0651  1.3617 -0.5379 -0.1512  0.3811 -0.8336 -0.8915 -1.4188
#>  1.4787  0.6774 -0.5002 -1.4241  1.3230 -0.5669 -0.2767  1.3881 -0.5382 -2.0356
#>  0.5402 -1.2473 -0.4384 -1.8250  0.2944  0.0776 -0.2196 -1.2268  1.0093  0.2477
#> -0.3358 -0.5431  0.4627 -0.8441  0.1997  0.5466 -0.9575 -1.4366 -0.7110 -0.4505
#> -0.1945  1.3131 -0.1930  0.3379  0.6802 -1.1062  0.7706  0.4912 -0.4113 -3.0000
#> -2.2572  0.0320  1.6174 -0.2960  0.5944  0.0129  0.0359 -0.4307 -0.7378  0.3461
#> -2.0619  0.9153  1.0512 -0.9741 -0.1762 -0.6839 -0.4871 -1.0355  0.9661 -0.8786
#>  0.3402 -0.2548  0.3919 -0.6501 -0.6694 -1.8570  0.1571  0.7122 -0.4892 -1.2091
#> -1.4804 -0.4599 -0.0726 -0.7728 -0.7210 -1.1677 -0.4139 -0.3292 -0.2412  0.3507
#>  0.7638  2.2292  0.0213  0.5032 -0.7512  1.6318 -1.2238  0.6215  0.9247 -0.0961
#> -0.9612  1.5517 -0.9174  1.1127 -1.5519  0.2099 -1.6870 -1.2619 -0.6493 -0.0362
#>  1.0402  0.1273 -0.5202  0.2896 -0.8741 -1.2843  1.2134 -0.2533  0.5004 -0.0222
#> -1.8151  1.0241  2.3155  1.3970 -0.2362 -0.2948  0.1317  1.7130  0.5065 -0.7321
#>  0.1926 -0.1383 -0.2410  0.9184  0.9431 -0.1219 -1.4058  0.6776 -0.9753  0.5968
#> -0.2829  1.4245 -1.1951 -1.1778 -0.8235 -1.7456 -0.0541 -2.4212 -0.6903  0.2007
#>  0.1952 -0.2090  1.4444 -0.1056 -0.4567 -0.4069  0.0401 -1.2537 -0.8961  1.4282
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]