Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.0579  1.4670  1.7213 -0.3225  1.0899 -0.2366 -0.2201 -0.4109 -1.1624  1.2072
#> -0.6591  2.3345 -0.3415 -0.0508 -0.8578  0.4786 -0.6148 -0.9858 -1.1743  1.0466
#>  0.6170 -0.9897 -0.0565  1.0330 -0.8760  0.4162  1.8492  0.8736  1.3621  0.5767
#>  0.4107  0.3889 -0.4802 -0.9153 -0.3160 -0.5975  0.1734  0.5409  1.7908 -0.0035
#>  0.5005  0.8967  1.0663 -0.8555  0.8941  1.2959 -0.4302  1.0882  1.4296 -1.1244
#> -2.4646 -1.1219  1.0192 -0.2042  0.4101  0.3261  0.0245  1.3646 -0.5277  0.3987
#> -0.0336  0.1663  1.4745  0.0978 -0.6565  1.2529 -0.3047  0.9786  0.2512 -1.2266
#>  0.3905  0.1101 -0.6882 -1.1749 -0.3265 -1.1287  0.1415 -0.6274  0.6135 -0.1059
#> -0.5529  0.7282 -0.2074  0.5303  1.3072 -1.8432  0.5747 -0.0582 -0.3669 -0.6630
#> -0.5391 -0.2200 -0.7846 -1.0483 -0.5114  0.2661 -0.5038  0.0146  0.6947 -0.2638
#> -0.7704 -1.3854 -1.0551  0.2371 -0.1996  1.5248  2.0299  0.0029 -0.5331 -2.2832
#> -0.0655 -1.1602 -0.1708  0.1443  0.5753  0.9510  0.6961  0.5192 -0.3866  0.6145
#>  0.2378  0.8304 -0.2884 -0.9319  0.0705 -0.4152 -1.0771  1.8018 -0.9671  1.3673
#> -0.7475 -2.0562 -1.5969  0.2981  0.2367 -0.1385  0.6716  1.0046 -0.7195 -2.0439
#> -0.0662  0.9243  0.7662 -0.2285 -0.8238  0.1889 -1.0375 -0.4806 -0.2393  1.1494
#>  0.1200 -1.0079 -1.1425  0.2435  0.7072  1.1711  0.0987 -0.6180 -2.2362  0.3744
#>  1.5221 -1.3383 -0.5972 -0.1737  0.2954 -0.1054 -1.6241  0.3687 -0.3378  0.3166
#>  0.7870 -0.5046 -0.5883  0.3342 -1.1596 -2.0607  0.0594 -1.0648  1.7060  1.3576
#>  1.6319  0.4473  0.0906  0.5603  0.7568 -0.0503 -2.1431  0.9372  0.1134 -0.2419
#>  0.7872  0.4526  1.3965  0.8918  0.3717 -2.0397 -0.9891 -0.3715 -0.1944 -0.3208
#> -0.0684  1.7176  0.5131 -0.6637 -1.9214  0.8431  1.7037 -1.2057 -0.9775  0.0716
#> -1.3856  0.6967 -0.0281 -0.3907  0.6452 -0.1677  0.8939  0.8608 -0.1921 -0.6805
#> -0.1975 -1.0141 -0.4169  0.5094  1.7315  0.7364 -1.1331  0.5738 -0.0868 -0.3955
#> -0.7449 -1.4545 -0.9715  1.0000 -0.7924 -1.0290  2.2427  0.0238  1.4558 -1.8374
#>  0.1820 -0.1592 -1.1705  0.0335 -0.8642  0.3354  0.9987  1.3475 -1.5973  0.1202
#>  1.6007 -0.5058 -0.8241 -0.5481 -0.4377 -1.7075  0.9758  0.7342 -1.1781  1.6533
#>  0.1397  0.8612 -1.7984 -0.3760  0.2177  0.3333  0.5610 -0.1014  0.8146 -1.6031
#>  0.2030 -0.2361  0.6154 -2.0449  0.6970 -1.1396  0.8685  3.1209  1.1718 -0.2872
#>  1.0258 -0.7328  0.1863  1.3876 -0.9023 -0.8717  0.0270 -0.0647  0.5476  0.6884
#>  0.7583 -0.4108  0.8115 -1.1130  0.8755  0.3764  0.5084  0.5622  0.7141 -2.9436
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]