Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.0442 -1.4195  0.4622  0.3812  0.8533  0.7282  0.1868  1.7330 -0.7899  0.5156
#>  0.1913  2.0218 -2.5383  0.5441 -1.3651  0.0548 -0.8014 -0.6535  0.7717 -0.9829
#> -0.2730  1.4975 -0.0068 -0.3847 -0.0528 -0.5433 -0.2145 -0.1337 -1.1503  0.8485
#>  0.9900  0.1189  1.0354 -1.1091  1.0489  1.0277  1.0116  0.2186  2.9895  0.6467
#>  0.3450 -1.2611 -0.1695  1.9535  1.8784 -0.7916  0.4136 -0.1196 -0.3446  0.8585
#> -0.2607 -0.4255 -1.1340 -1.2357  0.9362  0.2630  0.1362 -0.3544 -0.0999  0.8336
#> -0.2812 -0.6242 -0.7514 -1.6121  0.4269 -0.1543 -0.2352 -0.2199 -1.3546  0.3522
#> -1.3797  0.7052 -1.0575  1.1575  0.2025 -2.1325 -0.7351  0.6229 -1.0880 -0.1889
#> -0.6704 -1.7117 -0.2587 -0.1723  2.2755 -0.2005  0.1483 -1.1435 -0.4574 -0.4624
#> -1.6050  0.5435  0.5532  1.9916 -0.0208  0.6693  0.5416  1.1549  0.0990  0.4371
#> -1.9832  0.0303 -0.2387  0.5036 -0.1597 -0.6502  2.1141  0.0090  0.2328 -0.5541
#> -0.7914  0.8725  0.3067  0.7259  0.0373 -1.3419  1.0566  0.7187  0.8984 -1.6496
#> -1.7326 -1.7459  0.7743 -0.2226 -0.5126  1.3429  1.0583 -0.5698 -0.7487  1.0866
#> -0.6508  1.0654  0.3290 -0.7769 -1.4012 -0.9675  0.2821  0.4117 -1.3156 -1.1105
#>  0.9470 -0.4941 -0.6292 -0.0679 -0.5448 -0.0037  1.3622  0.2021 -0.3689 -2.3966
#> -1.1468 -0.1281  0.9782 -0.4902  0.1829 -0.1566  0.3493 -0.7050 -1.2766  0.8133
#> -0.0257 -0.4697 -0.1305  1.4548  0.1423 -1.2685 -0.3350  0.3550  0.2025 -0.6949
#> -0.7982 -0.5956  0.0731 -1.1380 -0.1070 -0.3082 -1.3779 -0.6213  0.7297  0.4703
#>  1.8193 -0.4542  2.0720  0.4018  0.1783 -0.3169  0.6525  1.7154  1.7538  0.5222
#> -0.0234 -0.1124  0.0656 -0.1956 -1.1114  0.3370  0.3787 -0.0964  0.7600 -0.7704
#> -0.5259  1.2241  0.9098  0.1067 -1.1837  1.1771  0.5579 -0.3608 -1.0047  0.3192
#> -0.4595  0.6784  1.3393 -1.8498 -0.0425 -2.0414 -0.3107 -0.2905  0.8001  0.4057
#> -2.9645 -0.1805 -1.0777 -0.6184 -0.3076  0.3170  0.2142 -1.2946  2.0190 -1.8219
#>  1.1878  0.0966 -0.8611  1.3476  0.0287  0.6398  0.8937 -1.5620 -0.5741 -0.8346
#> -0.4822 -1.0305  0.0371 -0.1172 -1.3774 -0.5622  0.7512 -1.5492 -1.2033  0.3319
#>  0.3152 -2.0310  0.7978  0.1351  0.6004  0.0797  0.9046  0.5536 -1.5747 -0.9064
#>  0.2363 -0.0532  0.6950  0.0161 -1.4893 -0.8564  1.1309 -1.0882 -1.3269  0.2377
#> -0.8692  2.1759 -0.2764 -1.0613 -2.0233  0.8437  0.2385  1.0377  0.6824 -0.1300
#>  0.4230  1.9074  2.0296  0.3149  0.6424 -2.1777  1.6976  0.3992 -0.8108 -0.5173
#>  0.8210  0.9385 -1.9629  0.7368 -0.4888  1.7569 -1.1523 -0.4409 -2.6393 -0.3387
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]