Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.0153  0.6121 -0.2358 -0.4208 -1.3985 -0.4734 -0.7837 -1.2355  1.5896 -0.6016
#>  2.4616 -0.9175 -1.3937  0.0475  2.2071  0.4515 -0.9222  0.8335 -2.5703 -1.1126
#>  0.8134 -1.1224 -0.7698 -1.1305 -0.8686 -1.2284  0.8622  0.4585 -0.9211 -0.4100
#>  0.7000  2.4647 -2.3151  0.5502 -1.1960  0.1825  0.6401 -0.9074  0.2670  0.0845
#> -0.6616  0.6904  0.3191 -0.5068 -0.9940 -1.0645 -0.0648 -1.1178 -0.8630  0.0309
#> -0.2091 -0.3044  0.8134  0.2097 -0.5613  0.1151  1.4666 -0.9293  0.9197 -0.4561
#> -0.0377 -0.6130 -0.5381  1.6968  0.5801 -0.5823 -1.0003 -1.0063 -1.1251  0.4064
#>  1.9085 -0.3446  1.4669 -1.4862  0.3186 -1.2384  1.2052  0.3101  1.2048  0.6496
#>  1.9063  1.5511  1.4929 -0.2947 -1.0654  0.1391  0.8904 -0.4436  0.8640  0.2857
#>  0.4892 -0.8232 -0.3455 -0.7817 -0.2176 -0.9214 -0.2314  0.7615  1.6163 -1.6061
#> -0.1961 -1.6242  0.6578  0.5667  0.1428 -1.4427 -0.8447  0.6239 -0.0754  1.4050
#>  0.4031 -0.2607  0.8615 -1.0353 -0.3449  0.6853 -0.4669 -0.9202 -0.9564 -1.1635
#>  0.2231 -0.1359 -0.4712 -2.2198 -0.3197  0.5140  0.4679  1.0145 -1.2089 -0.4686
#>  1.1496 -0.1169  0.4586  1.4487  1.7085 -1.4634 -0.2056  1.5130  0.7976  1.4648
#>  1.3653  0.9863  1.4818 -0.1155 -0.3669 -0.6308  1.0037 -0.3456 -0.5744  1.2764
#> -1.2241 -0.1591  0.3599  0.8733 -1.7754  0.1991 -0.6989  2.1301  0.2520  0.6688
#>  1.8454 -0.1352 -0.7778  0.0279 -0.7180 -0.6834 -0.7306  0.2617 -0.3870  1.1875
#>  0.0323  1.7788  1.4638 -0.0767 -0.1966 -0.4662  0.0117  1.9551 -0.3975  0.8222
#> -0.2616 -1.6651 -2.0178  0.2330 -0.8358  0.2146 -0.5113 -0.3623 -0.2526 -2.0426
#> -0.7613  2.0277  0.9099 -0.0002  0.8652 -0.9943 -1.6832 -2.2119 -0.2471 -1.9909
#> -0.4405 -1.8734 -1.2104 -0.8024 -0.1939  0.3298  0.4512 -0.6271  0.0455 -0.4470
#>  0.9558 -0.6140 -0.2158 -1.3350  1.0155  0.4592 -1.8118  1.6897 -1.5919 -0.5012
#>  0.3984 -0.1519 -1.1872 -2.0747 -0.8311 -1.6191 -0.5103  1.0422 -1.9656 -0.2401
#>  0.5166  0.5266 -0.3427  0.5402 -1.3067 -0.8624  0.7958 -0.5716 -0.0386  0.4802
#>  1.2428 -2.4123 -0.1214 -1.0890  0.6675  0.1202 -0.3488 -0.7908 -0.8703  0.4774
#>  0.6825  0.0809 -0.0443 -0.3639  1.6688  1.3193 -0.1920 -0.0373  2.2173  0.7068
#> -1.4824 -0.1229 -2.3431  0.3808 -1.1983  0.8033 -0.1542  0.5800  0.7800 -1.3108
#> -1.0054 -0.5866 -0.6478 -0.5916 -0.3294  0.3021 -0.3106 -1.5834 -1.0160 -0.1077
#>  0.8543 -0.3574 -0.9106 -0.4470  0.4715 -0.0294  1.0622 -1.1453  0.3791  1.0256
#> -0.7481 -1.9769 -0.2906 -1.2114 -1.0244 -0.4315 -0.0271  1.5009 -0.5404  0.1782
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]