Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 4.5220e-01  4.9635e-01 -1.1116e+00  3.6822e-01  2.2517e+00 -1.8780e-01
#> -3.8111e-01  3.3697e-01  8.7529e-01  2.0847e-01 -3.3850e+00 -3.8535e-01
#>  1.4936e-01  4.4151e-02  5.9584e-01  7.5620e-01  1.5094e-01 -4.4838e-01
#> -2.6955e-02  2.7105e-01  1.5332e+00  4.9919e-01 -1.2331e+00 -2.5926e+00
#>  9.0290e-01  1.1856e+00 -2.1853e+00  7.1565e-01 -4.0728e-01  6.5265e-01
#>  1.5032e+00 -4.0478e-01  1.5681e+00 -9.3325e-01  8.5149e-02 -8.3270e-01
#>  1.0551e+00  4.7361e-01  1.4728e-01  1.3353e+00  4.7373e-01 -5.5329e-01
#> -1.2637e+00 -1.2282e+00  1.4469e+00  7.1909e-04 -3.7302e-01 -1.4506e+00
#>  6.4746e-01  5.1958e-01  7.2038e-01 -3.1268e-01 -1.6649e+00  8.8828e-01
#>  3.5689e-02 -5.8741e-01 -4.1198e-01  3.3998e-01  1.0887e+00  3.9488e-01
#> -1.8322e-01 -6.3537e-01 -6.8920e-01  1.6011e+00  1.0528e+00 -1.3415e+00
#>  5.9472e-01 -7.6422e-02  3.6217e-01 -1.8371e+00 -1.0777e+00  5.6465e-01
#> -1.2401e+00 -1.7452e+00  9.1685e-01  7.7169e-01  1.5780e-01 -1.4562e+00
#> -1.9489e+00 -3.1748e+00  7.4021e-01  4.5366e-01  9.5403e-01 -3.3297e-01
#> -1.0507e+00  3.5319e-01 -1.5735e-02  1.8627e+00 -8.9403e-01  5.9830e-01
#>  2.8157e-01  1.2734e+00 -5.3873e-01  5.7123e-01  7.5107e-01 -1.8522e+00
#> -3.6125e-01 -3.0983e-01  1.0008e-01 -5.5861e-01  5.3324e-01  3.2186e-01
#> -7.7316e-01  1.2418e+00  4.5732e-01  2.9205e-01 -5.0979e-01 -5.4490e-01
#> -2.5350e+00  6.3680e-01 -7.7185e-01  3.1862e-01 -1.9686e+00  2.5230e-01
#> -1.5331e+00  2.2876e-01 -1.1277e+00  9.6517e-01 -2.3886e+00  1.9645e+00
#>  4.2324e-01 -1.4521e+00  1.2146e+00 -2.3027e+00 -9.6881e-01 -8.2591e-01
#> -6.0432e-01  3.7716e-01  8.3517e-01 -1.1814e-01 -1.4696e+00  8.4752e-01
#>  1.8076e+00  6.8342e-01 -4.4359e-01  9.2820e-01 -1.5695e+00 -1.3953e+00
#>  6.2006e-01  4.7680e-01  1.0912e+00 -5.1365e-01  3.0998e-01 -1.3237e+00
#>  1.8413e-02 -1.1150e-01 -1.8252e-01  1.2104e+00 -6.1832e-01 -1.1692e+00
#>  4.0440e-01  9.7780e-01 -3.2511e-01 -8.0893e-01 -7.3999e-01 -6.5052e-02
#>  1.7183e-01 -5.1699e-01 -2.2971e+00 -3.4342e-01  1.6791e+00  3.8392e-01
#> -1.2727e-01  6.6432e-01 -1.3852e-01  2.8316e-01  4.9306e-01 -1.0742e-01
#> -6.4500e-01  5.1774e-05  1.2519e+00  2.6521e-01  5.3223e-01  1.8048e+00
#> -2.2939e-01 -1.4549e-01 -8.5208e-01  4.8248e-01  2.7059e-01 -7.4894e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]