Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.3580  0.1383 -1.2673  0.2501 -0.2841  1.4166  0.6282  0.5389  0.0944 -0.3931
#> -1.2107 -1.1567 -0.7038  0.5264 -0.6398 -0.1027 -1.9397  0.3954 -0.8696  0.6584
#> -0.1280  0.0677  0.2582  0.4206  0.6341 -1.5423 -0.9506 -0.3545 -1.0648 -1.6256
#>  1.3722 -0.6718 -1.9431  0.3307  0.3537 -0.2980  1.0238  0.3549  0.8585 -0.4159
#>  0.7801  0.9949  0.7356  0.2714 -0.1072 -0.5425 -0.7610 -1.0079  1.1961 -0.1804
#>  0.6656 -1.4804 -0.0121  0.1893 -0.3458 -0.5814  2.2690 -0.5057  1.1068  0.6836
#>  0.7605  0.0264 -0.3774  1.4779  1.6009  0.5108  1.9417  1.5664  2.6962 -0.9384
#> -0.1347  0.8804  1.1932 -0.8758  0.1869  0.0672  1.1743  0.0451 -0.0136 -1.0161
#>  1.1854  0.9558  0.9469 -1.2606 -0.1517 -0.0151  3.2680  1.6308  0.5627 -0.6350
#>  0.7939 -0.1875  0.0479  0.6429 -1.0896  1.5446 -0.0932 -0.5807  1.8878  1.0400
#>  0.5652  0.1618  0.6693 -0.5079 -0.1813  1.3056 -0.7466  1.1954  0.2254 -1.0955
#>  1.4089 -2.1634  0.9012 -0.1834  1.2105 -2.8204  1.8618 -0.9073  0.5698  0.2109
#>  0.0784 -1.0439  0.4140 -0.7378 -1.0681  0.0051  2.4260  0.4637  1.7970  0.1388
#> -1.2694  0.8560  0.0118 -0.7514  1.4483  1.4956 -1.1083 -1.4732 -1.6527  1.4176
#> -0.4381  0.2222  0.3586  0.1748  0.4219  1.2145  0.4492  0.1128  0.8096  0.6340
#> -1.0861  0.0076  1.5809  0.0654  0.6439  1.6983  0.0667  1.6083 -1.7580  1.1033
#>  0.8410  0.0103  0.5820  0.9323  2.0109 -0.1583 -0.7921  0.6776  1.9564  0.5242
#>  1.4189 -1.3379  0.5481 -1.7782 -1.1976  1.4828 -1.5664  0.9899  1.8295  0.8446
#>  0.2474 -1.7510 -0.3433 -0.1640 -0.2122 -0.5932 -0.5442 -0.4568 -0.3640 -0.6401
#> -1.7952 -0.6114 -1.2079 -0.9159 -1.0374  1.1891  0.4260 -0.2923 -0.0211  0.2068
#> -0.8840  1.7064 -0.0484 -0.5424  0.7436  1.0007 -0.1824  0.3564  1.0609 -1.1968
#> -0.0290 -0.9936  0.0626 -0.6629  0.4851  1.8279 -1.2584 -0.6853 -1.0668  0.8362
#> -0.7535  1.9660  1.0958 -0.3363 -0.5292  0.2865 -1.2890  1.0470  0.0071  0.0728
#> -0.2140  0.5057 -0.4709  0.2200  0.1017  0.0149  0.1542 -2.3002 -0.8158  0.5106
#>  1.6554  1.2890  0.4219 -0.0415  0.0329  0.5018  0.1438 -1.7585  0.0277  1.0559
#>  1.8431 -1.1735  0.8418 -0.8727 -0.2184 -0.5784  1.8078 -1.4501  0.0588  0.1556
#> -0.7088  0.9606  0.3878  1.4544  0.4465 -0.5879  0.1373  0.6865 -0.3508  0.3515
#>  0.6804 -0.1177 -2.1129  0.4796  2.2838  0.9173 -0.8375 -1.9737  0.2780 -0.0691
#>  1.2184  0.0661 -0.1825  0.8086  1.3986 -0.1551 -0.0101 -0.5347 -0.2063  0.6542
#>  0.6704 -0.4774 -1.9621  1.3136 -0.3972 -0.0613  0.8505  0.7489 -1.7455 -0.4683
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]