Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.3356  0.2777  0.0855 -0.1872  0.2606  0.3183  2.1265 -0.6978  1.0651  0.8354
#> -0.8786 -1.0873 -2.4412  0.4060 -1.7839  0.9226 -0.4944  2.0245  0.7919 -0.3973
#>  0.6631 -0.4966  1.1780 -1.4742  0.5336  0.4590 -0.7922  0.0115 -0.1934 -0.8128
#> -1.2031 -0.2209 -0.8010  1.1201 -0.2978 -1.5737  0.1137 -1.8100  0.5998 -1.3275
#> -0.3108  0.5417  0.1368 -1.1715  0.8500 -0.4348  0.2261 -0.1143  0.2679  0.3966
#> -1.4708  0.9787 -0.2468 -1.5041 -1.4149 -0.5580 -0.6779  2.0564 -1.1662  1.2692
#>  0.6640  0.2233 -0.8260 -1.1259  0.3438  1.5559 -0.2315  0.3062  0.9079 -1.4049
#>  0.7555  0.0099  1.0775 -1.5716 -0.0098 -0.2222  0.7779  4.3385  0.6361 -0.5455
#> -0.1482 -2.1817 -1.8767  1.1288  0.6689 -0.1018 -0.6306 -0.7000  0.8037  0.9790
#> -0.8284 -0.5328  0.8096  0.5977 -1.0475 -0.4183 -1.4620 -2.0090  0.3318 -2.2976
#> -0.5165 -0.8008 -0.7372  0.6957 -0.2019 -1.5833  1.9373 -0.1068  1.2386  0.1104
#>  1.1636  1.1097 -0.5660  0.2630 -0.4789  1.0936  0.1582 -0.2523 -0.5650 -0.0623
#>  0.4165 -1.1204  0.2864 -0.7584  1.5786  0.2261  0.1901 -1.2418 -0.4983 -0.0401
#>  0.6401 -0.9216 -0.2603 -0.4442  0.6736  0.2234 -1.3201  0.1776  0.5810 -0.6169
#> -1.7908 -0.4206 -1.6338  0.5458  0.1151  1.7636  0.9347  1.6349 -1.1616  0.4295
#>  0.1325 -1.0651  1.2547 -0.2830 -3.0017  0.6690  1.6762 -0.2874  0.6519  0.1851
#>  1.3252 -1.2875  0.2867 -1.1912  1.3317  1.6767  1.3441  1.4648  0.0789 -0.6278
#> -0.3163 -2.2657  1.3299 -0.6801 -0.0638  1.2739 -0.1311 -0.2491  1.8889 -0.9709
#> -0.9180  1.3832  0.2754 -0.0083 -1.3339  0.6619  0.0251  1.4082  1.2618  0.4908
#>  0.2847 -0.2155  0.5176 -0.1952 -0.1829  0.9406 -0.3667 -1.9381  0.5107  0.3394
#>  0.5600  0.5240  0.4008 -0.4295  0.3227 -0.8730 -0.9019 -0.4108  0.2787  1.2307
#> -0.6243 -0.3804 -1.1785 -0.8838  0.5065 -1.4703 -1.5100 -0.5297  0.0279  0.3546
#> -0.1352 -1.1048 -0.8689 -0.7531  0.6130 -0.1735 -1.1600  0.5362  0.3472  0.7064
#>  0.2501  0.7543  1.8681 -0.1491 -0.4303  0.9933 -2.3398  0.0746 -0.9018 -0.0768
#> -1.2873 -0.2799  0.6229  0.1184 -1.2952 -0.7533 -0.7287  0.1584  0.3799 -1.9895
#> -0.1318  0.6462 -0.9084 -0.0387 -0.0224  1.6573  0.5079 -1.1896  0.5116 -0.4463
#>  0.5675 -0.1533 -0.5382  0.6943 -0.4453 -0.0116 -0.9380 -0.0128 -0.0023 -0.9179
#>  0.6504  0.8380 -0.8557 -0.3157 -0.1880 -0.3934  1.3584  0.0193 -0.1883 -0.1300
#> -0.1685 -0.5689 -1.3288 -1.5634  0.7791 -0.4159 -1.5910 -1.6147  0.6131  1.6020
#> -0.3716 -1.2786 -0.4270 -1.7035  1.1973 -0.1239  0.5286  1.3785  0.1390 -1.4170
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]