Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5858 -1.2279 -0.4402  1.2102 -1.6778 -0.6399 -0.7615 -0.3136  0.2912 -0.4176
#>  0.6708  2.0204  0.7981  0.0223  0.6907 -2.7920  1.4809 -0.6905 -0.7207  1.4340
#> -0.4881  0.3735  1.2766 -0.5138 -0.1219 -1.7487  0.2709 -0.5935 -3.1811 -1.1653
#> -0.3578 -0.5920 -0.9199  0.1137  2.1882  0.2001 -0.1881 -0.1617 -0.6975 -0.1060
#>  0.2093  0.1580 -1.6723  0.3150 -0.7504  1.5069 -0.9349 -0.4175  1.7328 -0.5394
#>  0.7092 -0.7950  3.6558 -0.7175  2.4479 -0.6031 -0.6409 -0.9297 -0.4507 -1.1258
#>  0.0006  0.3689 -1.0175 -0.4388 -2.0024 -0.3447 -0.4106 -0.2623 -0.6737  1.0769
#> -0.6796  0.0090  0.6209  0.5340 -0.1954 -0.6926 -0.3107  0.0575 -0.6258  0.2513
#> -1.2318 -0.7062 -0.7544 -0.3575  1.0088 -0.6485 -0.4551  1.0544 -2.1999  0.5051
#>  0.4025  1.1761  0.7897  3.0868 -0.9291 -0.4310 -0.4293  0.1871 -0.2898 -0.3262
#> -0.0162  0.0984  0.8871 -0.3665  0.3781  0.2657 -1.5230 -1.7600 -0.4452 -1.0615
#> -1.2378 -0.5519  1.1802  0.8453  0.7647  3.0179 -0.8007  0.0074 -0.1574 -0.8219
#>  0.0048 -1.3917  1.9815 -0.4628 -1.2116  2.7241  1.0624  0.4288  0.1065 -0.9768
#>  1.5587  0.3763 -2.2310 -1.0012  0.4069  0.3278 -0.5272 -0.4526 -1.2510 -0.8541
#>  0.4055 -0.6243 -0.5928 -1.6421 -0.0966 -0.9102  0.7990 -1.0297  0.8059 -0.8486
#> -1.8961  0.2518  0.2239 -0.1119 -0.7374 -1.1412  1.6084 -0.1660 -1.2716 -0.2404
#>  0.4016  1.0806 -0.5386  0.9698  1.5437 -1.0274  0.8015  1.0796  0.1077  0.8187
#> -0.2411  0.3774  1.2038  0.3059  0.2011 -0.4052 -0.7879 -0.0079  0.6278 -1.7954
#> -0.2262  0.4794  0.5906  1.5610 -0.2780  0.0331 -0.9908 -1.4527  1.3546 -1.5327
#>  0.1879  1.1092 -1.3004 -0.7274  0.7240 -0.3969 -0.4269  0.7149 -0.5201 -0.2569
#> -1.5270 -0.0150  0.3707  1.0542  1.5702  1.3818 -0.5810 -1.3789 -1.0625  1.1941
#> -0.6486 -0.0109  0.1316  1.9515  0.4003  0.4771 -1.8644  0.4847  0.3557 -1.6884
#>  0.0732  0.8982  0.3565  0.4702  0.1331 -0.7243  0.9979  0.0281 -1.8172 -0.1063
#>  1.0902 -0.6723 -0.4347  0.1537 -0.0885  0.0802  0.2753  1.2420  2.4538 -0.3722
#>  0.7489  1.6262 -0.1405 -1.9831 -0.7633 -0.7069  0.7753 -1.2698  0.9621  0.3238
#>  1.3163 -0.8955  0.6901  0.9719 -1.4675 -1.0510 -0.4380  0.2070  0.0576  0.5047
#>  0.4731 -1.0737 -0.7494 -0.1089 -0.2147  0.2282  0.6633  1.3145  0.1632 -0.2471
#>  0.1627 -0.3502  1.1058 -3.3060  0.2895  0.0458 -0.1651  0.4250  0.6917 -0.8982
#>  0.7182 -0.2118  1.1868 -1.3761  0.5977  0.9117  0.2892  1.4898  2.4148 -0.5403
#> -0.0488  1.1066  1.4122 -1.5116  0.4671  0.5948 -1.1412 -1.0091 -0.1120 -0.9543
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]