Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.8791 -1.4451  2.1480  0.8644 -1.3980 -2.4418  0.0295 -0.2297  2.7449 -0.1654
#>  0.9806 -0.7779 -1.9726 -2.0263 -2.4548 -0.2421 -0.3883  0.3490  0.8287  0.3408
#> -0.9816  0.0115  1.2784  0.3358  1.5643 -1.2268 -0.1834  1.8441  0.5409  0.6611
#>  0.4551 -0.8899  0.8084 -0.5889  1.3102  0.1785 -0.5397  0.5062  0.4081 -0.1021
#> -0.0898  1.1637 -0.4930  0.4557 -0.4362  1.6837 -1.0430 -0.7437  0.7854  0.6846
#> -0.8536  0.7292  0.7971 -0.2150 -0.4162 -0.3963  0.8324  0.5439 -2.0948  0.7070
#> -0.6843 -0.0907  0.2109 -0.2064  0.2099 -1.0917 -0.2540  0.1829  2.0042 -0.1559
#> -1.0975 -0.0447 -0.8114 -1.4194 -1.4773  0.2384 -1.1525  0.3142  0.7369  0.2193
#>  0.5710 -0.4618 -0.5571 -0.6098  0.7638 -2.2389  0.1897  1.5224 -1.3358  0.4222
#> -1.0104 -0.3663 -2.0090  1.1472 -2.8679 -0.3099  2.1548 -0.7839 -0.6251 -1.4005
#> -0.1304 -1.0361 -0.5854  0.4316 -0.3665  0.2821 -3.3803  0.7453  0.0350 -0.1470
#>  0.0130  0.8801  0.6001 -0.8080  1.6623  0.6034 -0.4892  0.6352  0.8170 -0.7039
#>  0.5760 -2.1411  2.0524 -0.0583 -0.9451 -0.4545 -0.1556 -1.0956 -2.5166  0.7560
#> -0.7339 -1.8531 -1.1023 -0.4241 -0.0398  1.5849 -0.6219 -0.0325  0.1329 -1.4808
#>  1.1425  0.9828 -0.1289  0.2134  0.9102 -1.3090  0.0793 -0.5618  0.3013 -1.6237
#>  0.3059  1.5933  0.8560  2.4311  0.3589  1.2483 -0.8887  0.1034 -1.2001  0.5388
#> -1.7678  1.9621 -0.5052  0.4910 -0.5173 -0.0408 -0.9682  0.6715  0.5180 -1.8687
#>  1.4070  0.2580  0.5108 -1.9010 -2.0296  0.5338  0.9981  0.9205 -0.1968  0.1234
#>  0.3016  2.3718  0.0556  0.1678  0.7490  1.1332 -0.2329 -0.7029  0.5457 -0.3549
#>  1.3090 -0.6544  0.5904 -0.3902  1.5430 -0.4779 -0.9755 -1.0771 -0.1369 -0.1480
#>  0.2314  0.8118  1.2661  0.3512  0.4813  0.4768 -0.1374  0.5168 -1.1957  0.3737
#>  0.0285 -0.5343  0.0511  0.3081  0.4004  0.7440  1.1902  0.3810 -0.5986  1.3737
#> -0.4057 -0.0183 -1.9203  0.7416  1.2392  1.5212 -0.0856  0.9715  0.0054  1.2100
#>  0.2816 -0.0932  0.3167  0.7377 -0.5146 -0.3968  1.0114  0.3769 -0.1446  0.4874
#>  0.3982 -0.0228  1.6541 -0.6015  0.9074  0.0976 -0.9691 -1.6350  1.3593  1.2491
#> -0.2919 -0.7000 -0.4166 -0.3059  1.2713  0.4295 -1.0117 -1.5303 -1.8389 -1.7526
#> -1.0300  2.0569 -0.3266 -0.1839  0.5338  1.2185 -0.7794 -0.1445  0.9649 -0.1859
#>  0.6875  0.3538 -0.2517  0.7239 -1.2716 -0.4131 -0.8790 -0.9192  0.5638  0.0102
#> -0.0298  0.8289  0.5545  0.7534  0.3348  0.2769  0.1581  1.4526  1.3528  1.2628
#> -0.8449 -0.1728  0.1832 -1.5629  1.3294  0.5122 -0.0004 -2.0546  1.8662  1.2404
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]