Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.1569  0.0241 -2.0610 -0.6264  0.5981  1.3125 -0.2388 -1.2286 -0.1043  0.6096
#>  1.1409  1.4804  0.1845 -1.2451  0.6974  0.7438  0.2576 -0.3703 -1.0992 -1.1761
#> -0.2716  0.7948 -0.8931  1.2814  0.1433  0.2315  0.3704  0.2461  0.7108  0.1342
#>  1.4553 -0.5825  0.4021  0.6264  0.6403  0.5774 -1.7916  0.9597 -1.3907  1.8177
#>  0.9435  0.9173  0.2952  0.5838 -0.1443 -0.6109 -0.5650 -0.3410  0.9112  1.3965
#>  0.1802  0.5131  1.8577 -0.7561  0.3995  0.7962  0.6359 -0.2854 -0.7033  3.1278
#>  0.5228  1.0876  1.0990  0.0375  0.8070 -1.0008 -0.4575  2.0482 -0.1919  0.2672
#>  1.8834  0.2107  0.2717 -0.3967  0.6948  0.8214 -0.7907 -2.0670 -0.2719  0.0449
#>  0.0707  0.8241 -0.2036  0.0607 -2.2944  0.6977  1.2207 -1.0731  0.3186 -0.0272
#>  0.2853 -0.4095 -0.1054  1.3285  0.0422 -0.8792 -0.7862 -0.6847  1.2981 -1.1092
#>  0.3589  0.8222 -0.0766 -0.5896 -0.6205  0.6897 -0.5921  1.8778  0.9070  0.4690
#>  0.1213  0.5302 -0.6835 -1.0592 -1.2316 -0.2684  0.4636  0.1665 -1.0632  0.8463
#> -0.6995  0.7486 -0.6751 -0.3645 -0.2682 -0.4095 -0.3072  0.5380 -0.0660  0.7635
#> -1.0437  1.3592  0.5481  0.3979 -0.1412  0.3833  0.4515 -0.4272 -0.6917  0.0536
#> -0.3673 -1.3163 -0.7214 -2.6054 -1.2276 -0.5706 -0.1930 -0.8293  0.1672 -0.5191
#> -1.4556 -0.6876  1.3984 -1.6435  0.0011  0.5767 -1.0178  0.2638 -0.1431 -1.6479
#>  0.3163 -0.9361 -0.5896  1.1044 -1.2941 -1.2151 -0.7215 -0.2395 -0.2836 -0.1467
#>  1.4294 -1.3212  0.2351  0.6179 -0.6063  0.2090  0.0380 -1.0306 -1.0569  1.0985
#> -0.7003  1.0848 -0.6361 -0.7942  0.1213 -0.9164  1.5910 -0.2325 -1.1302  0.6234
#> -0.9586  0.5615  0.1785  0.1083 -0.1544  0.0556  0.6548  1.8075  0.0431 -0.0160
#>  0.1125 -0.7503 -1.2616 -0.2730 -0.2655 -0.2539  2.5935  2.5553 -0.2149  0.6272
#>  1.9037  0.9864  0.4536 -0.5498 -1.1576  2.7704  0.7190  0.7096  0.4166 -0.1935
#>  0.6509 -0.9499 -0.8705 -0.4440 -0.2580 -0.4095  0.2310  1.0000  1.9094  1.6564
#> -2.3569  0.1108  2.0605  1.5905  1.7972  0.8360  0.1311 -0.2145 -0.7545  1.3264
#> -0.7295 -0.7618  1.1414 -0.3349  0.3552 -0.4016  0.4652  1.7565 -0.8263  0.0739
#> -0.1382  1.7322  0.2171  0.8837 -0.1343 -1.5038 -1.0946  0.1533 -0.2432 -0.9637
#> -0.9862 -1.2345 -1.6182  0.0684  0.6900  1.6136 -0.4302  1.1944  1.7988 -0.7202
#> -0.5323 -0.9074  0.7028 -1.1417  0.1007  2.4818 -0.2094 -0.0174  0.2763 -0.5279
#> -1.0083  0.0723 -1.4699  0.8882 -0.1720  0.4433 -0.0410 -0.5518 -1.3387 -0.9831
#> -0.6685 -0.6589  0.2190  1.1053  1.2035 -0.1396  0.5307  0.2488  1.3246 -0.6183
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]