Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.8196  0.5094  1.1729  0.6838 -0.5213 -0.1251 -0.4840  1.2976  2.2351 -0.6608
#>  1.8712  1.1299  0.6582 -0.3490  0.6932 -0.7390 -0.6569 -0.3548  0.8662  1.0500
#> -2.0573  1.1687  0.0261  1.0895  1.3505 -0.1174  0.0963 -0.0639 -0.4610 -0.1769
#> -1.0476 -1.8571  0.3645  0.2353 -1.2607  0.1815 -1.2922  0.8279 -0.0141  1.0029
#> -0.3035  3.5348  0.6360  0.7740  2.2463  0.2389 -0.2380 -0.3510  1.9770 -0.5988
#> -0.7183 -0.1410  0.8969 -1.1560  0.0416  0.0671  0.8191  0.6762 -0.4591 -1.0590
#> -1.4685 -1.2553  0.2580 -0.3334  0.2359 -1.6571  0.0375 -0.2607  2.0134  0.2233
#>  1.3296 -1.4923  1.1808 -1.2917  0.4408  1.6238 -0.1064  1.6166 -0.5151 -0.6122
#>  1.0411 -1.8738  0.9696 -0.2565 -0.1549  1.4315  0.5903 -1.7338  0.1699 -0.6690
#> -0.5584 -0.5649 -1.7336  0.0593 -1.1803 -0.1201  1.6017 -0.9727 -1.8760 -0.2694
#> -1.2843  0.2477 -0.0037  0.1374  0.9908  0.3645  0.4154  1.3335 -0.2088 -0.9682
#>  0.9053 -0.6179 -1.8166  1.1653  2.2359 -0.9190  0.4064 -1.2196 -0.3930 -0.6228
#>  1.0496 -1.5178  1.0008  0.3591  0.9278 -1.7733  0.8127  0.7526 -1.2933  0.1173
#>  1.1927  1.2571  0.6359 -0.0473 -0.6400 -0.6679  1.8041  1.2144  0.1552  0.4803
#> -0.0243 -0.7650  0.2207  1.8203 -0.4085 -1.0985 -0.1206 -0.9662 -0.2399  1.3854
#>  0.8513  1.0701 -1.3896  1.4648 -0.7346  1.2433 -1.4831 -0.9263  1.1742 -0.5611
#>  0.0683 -1.1151 -0.4451 -0.7778 -0.1570  0.7447  0.9085  0.2056  0.0838  0.4049
#>  1.2908 -1.7980 -0.1069 -0.0348 -0.1200 -1.0570 -0.9873  0.6986 -0.3076  1.1131
#>  0.4248 -1.1851 -2.3685 -0.2176  0.5166 -1.7953  1.6334 -0.2082 -0.0485 -0.9526
#> -1.3860 -0.7927  2.6538 -0.0800  0.5057 -0.9404  0.7995 -0.1510  0.4879 -0.3651
#> -0.3173 -0.8305  0.7717  1.6874  0.8836 -0.9079  0.2803 -0.9654 -2.9379  1.8430
#>  0.1144 -0.7003  1.2769 -2.1335  1.0140  1.0321 -0.4514  1.3916 -0.6621 -1.2480
#> -0.1182  0.8563 -0.0198 -0.7370 -0.4523 -0.2007 -0.1333 -0.3524  0.0832 -1.1519
#> -0.9387  0.2505  1.4121 -0.6058 -1.3575 -0.4983  0.9496 -2.1535  1.3819  1.3805
#> -1.7636 -0.0877  1.7952 -0.0200 -1.8510 -1.6686 -0.5237 -0.5963 -0.3772  0.7214
#> -0.9267  0.6513  1.4309  0.4110 -2.8307  1.7369  1.8541 -0.2024 -0.0862 -0.1319
#>  0.9494  1.2028  0.5471 -1.4686 -0.2667  0.0869  0.6988 -0.7071 -1.1386 -0.5930
#> -0.0924 -0.7891  0.2318 -0.3189 -1.9285  0.5501  0.5835 -1.1030  0.1290 -0.8012
#>  0.9269 -0.5623 -2.4030  0.2092  0.9451  0.2561  1.4523 -0.2036  0.8836  0.5513
#> -0.7822  0.4472  0.0334 -0.6966 -2.8408  0.5164 -0.9280 -2.2593  0.2858 -1.5856
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]