Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6040  0.5936  1.1917 -0.0741 -0.4838  0.2996  0.6644  0.2155  0.4235  0.6301
#> -0.9014  1.5024  0.7035  0.6862 -1.0436  0.1912 -0.7983  0.1784  0.4130  1.9249
#>  0.2846 -0.9278 -0.2017 -0.5959  1.6210  0.9948  0.1803 -0.3920 -0.3355 -0.0872
#> -0.3844  0.3788  0.1911  0.0611 -0.3838  0.2279  0.4222 -0.5350  0.3187 -0.6018
#> -0.8673  2.0718  0.8761  1.6511  0.3068 -0.2563 -0.6175  0.7167  2.3454 -0.9820
#>  0.6065  0.8052  0.5234  0.6148 -1.2417  0.2578 -0.8725 -0.0082 -0.3416  0.4022
#>  0.8772  0.0803 -1.1336  1.0805 -1.9125  1.0008 -0.5236 -0.8234 -1.7397 -0.5591
#> -0.5943  0.1516  1.3381  0.0648  0.5297 -1.0554  0.9086  0.8161 -1.8713 -0.0344
#>  0.9722 -0.7379  1.2320  0.4624  0.1473  0.9010 -1.5709 -0.3389  0.4909  0.9039
#>  0.0030 -0.5274 -1.8579  0.9595 -0.5197  0.5671  0.4675  2.4144 -0.1652  0.3983
#> -0.9917  0.4434 -0.3309 -0.4404  0.5053 -0.1997 -1.0116 -1.5880 -0.2959 -0.7394
#> -1.1873  2.1147 -0.0420  0.7248 -0.4076  1.3777  0.7086  1.1235  0.0826  0.4270
#>  0.9474  2.1558 -0.3792  1.2309  0.9984 -1.2284  0.6360  0.1148 -0.0372  0.7222
#> -0.3878  1.4711  0.5015  2.2293 -0.0907 -0.0420  0.6198  0.6170 -0.5613  1.6637
#> -0.1778  1.1771 -0.1471  1.7177 -1.1260 -0.9468  0.3302  2.0033 -1.6160  0.5747
#>  0.1486 -1.4179  0.0829 -0.6000 -0.1754  0.2614  0.2433  0.6885 -0.5830  1.0387
#>  1.0241  1.6052  1.1382  0.4628 -2.2073  0.9262 -0.2300  0.0544 -0.2948  1.5343
#> -0.0447 -0.4109 -0.1430  0.5696  2.7461 -0.5267  0.3501 -0.2923  0.1445  0.3236
#> -0.7124  0.4380 -0.9000  1.8724 -0.8748  0.8262  0.6334  0.1635 -1.1822 -0.9693
#> -0.3841 -0.6580  0.1021  0.4545  0.2772  1.1633  0.0734 -1.0249  0.7943 -2.0865
#>  0.0881 -1.2007 -0.3599 -0.6833 -2.0719  0.4338  0.2653 -0.0455  1.1518  0.6594
#> -0.6161 -0.1289  0.7793 -1.5736  1.1453  0.7578  1.7775  1.3513  0.2847  0.7972
#> -0.4880  1.1981 -1.1831  2.6077 -0.3389  0.1451  0.8298  0.1293  0.0866  1.4862
#>  0.2906 -0.9075  0.2546 -1.6072 -1.8088 -1.3975 -0.4340  1.4509  0.9958  0.0933
#> -0.9046 -0.1371  1.1350  1.8749 -0.1219  0.0705  0.1512 -1.4868  0.0611 -0.1628
#>  2.5570  1.1702  1.3636 -1.5153 -0.2528 -0.5854 -0.1467  0.8987  1.0428 -0.7048
#>  0.8073  0.1540 -0.2344 -0.5583  0.6281 -0.7992  0.9735 -0.0668 -0.3894  0.6302
#>  0.3010  2.1264 -1.6806 -0.3383 -0.7240  0.5704 -0.5541 -0.4078 -0.1321  0.6102
#>  1.5639  0.8525  0.3130  1.2618  0.9485  0.6634  0.1553 -1.0162 -0.2551 -0.6110
#> -0.1631 -1.3213  1.1285 -0.2404  0.4250  0.1483  0.2092 -0.6864  1.2119  1.1861
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]