Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.9845  1.6729  1.1151 -0.2605 -1.1576  2.1440  1.9895  0.7944 -0.0459  0.4570
#> -0.2700  0.9442 -1.1667  1.9390 -1.5184 -0.6741  1.1598  0.5114 -0.6353 -1.5524
#>  0.6287 -1.8385  1.0687  1.3280  1.1388  0.5124  0.2004  0.3878  0.9016 -0.3468
#> -1.0941  0.4415  1.8348 -0.2903  0.3171  0.4101 -0.4741  1.7993 -0.8076  0.7616
#> -0.2105 -1.7707 -0.7675 -0.7274 -0.1936  0.3530 -0.9206  1.3738 -0.7472  0.4059
#> -1.3011 -0.0804 -0.2465  0.5668 -0.9299  1.2034  0.8746 -0.6702  0.3475  0.5043
#>  0.1556 -0.7141 -0.7615 -0.6150  0.8853 -1.3545  0.3867 -0.9911 -0.3307 -1.4074
#> -0.3670  2.5800 -0.2426 -1.7215  0.6391 -0.2448 -0.6396  0.5255 -1.4934 -1.6060
#> -1.4275 -0.5541 -0.4788  0.2306  0.1893 -0.1216  1.4296 -0.1781  0.1315 -0.8461
#> -0.8274  0.1195  0.4905  0.0519 -0.8943 -1.6175 -1.4693  0.7682  0.2181 -1.1072
#>  0.0626  0.9759 -0.0209 -1.4506 -0.1654 -0.6127 -0.8195  0.5984  1.2256 -0.0111
#>  0.7160 -2.5047 -0.5562  0.9171 -1.3705 -0.5299 -0.2642  0.2027  1.3689 -0.3792
#>  0.4468  0.2491 -1.3027 -0.9679 -0.8085  0.5136 -0.4729  0.4181 -0.7526 -0.4202
#>  0.8078  1.7366 -1.2348 -0.6268 -0.0098  0.7435  0.1140  0.3196 -1.0932 -1.4401
#>  0.9544  0.4687  2.1417 -0.5232  0.7540  0.7105 -0.5719 -0.4704 -1.1421  0.8966
#> -0.6350 -0.8704  0.5844  0.0713 -0.6548 -1.9892  0.5129 -0.8493  0.2705  0.1042
#> -0.3529 -0.1702  0.0743  1.2966 -1.2404 -0.0348  0.0732 -0.6620 -1.1024 -0.8601
#>  1.0207 -0.1061 -0.8435  1.0000 -0.6492 -1.8050  0.2917  0.2821  1.1202 -1.8296
#>  0.3495  0.0542 -1.2521  1.8833 -1.2097  0.9512  1.1726 -1.4477 -1.8503 -0.8723
#>  0.7578  0.4630 -0.3649  0.6253 -1.0982 -0.0271 -0.8321  0.8235 -0.3767  0.4934
#> -0.8259 -0.4342  0.0780  1.5406 -0.0507  0.0466 -0.6041 -0.3537  0.5886 -0.9722
#>  0.6489 -0.7591  0.8479 -2.0229 -0.1018  1.5471  1.0780 -1.1728 -0.7805 -0.6103
#>  1.0647  0.4495  0.9410 -0.9531 -0.1962  1.0232 -0.1025  0.7562 -0.7076  0.3639
#> -0.6236  0.0406  0.6897 -0.7674 -0.2960  0.5623 -0.5955 -0.3346  0.6311 -0.7773
#> -0.5530  1.5734  0.5291  0.9112  0.9110  0.3134 -0.8092  0.3853 -0.6244 -2.0222
#>  0.8437  0.4809  0.4011 -0.3022 -0.5814 -0.3383  0.4251  0.5903 -0.3459 -1.3330
#> -0.8288  0.8155  0.8643  1.2886  1.2965 -0.0812 -1.0308  0.5553  1.3919  0.5411
#>  1.3879 -0.0974  0.5489 -0.8345 -0.2008  1.1280 -0.9060  0.5886 -1.4510 -0.8494
#> -0.5709  1.8299  0.3870 -0.6994 -1.7368 -0.6300  0.4845 -1.6863  0.2873 -0.6325
#>  0.6304  0.4761  1.0880 -0.9705 -0.0368 -1.1900 -1.2529 -0.9086  1.3299 -0.5443
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]