Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.2223 -0.9928 -0.7325  0.8157  1.0704 -1.5257 -1.3454  0.1028 -1.8389 -1.1203
#>  0.5403 -0.0846 -1.2236  0.5108 -1.3609  1.0243  0.1317  1.3702  1.0619  0.7497
#> -0.1832 -0.1175 -0.2759  1.0349 -0.4090  1.3167 -1.4790  0.2468  1.0417 -0.4322
#>  0.0297 -0.0476 -0.5310 -0.5005 -0.0121  1.7166 -0.6547  0.3166 -0.0751  0.3239
#> -0.3730  0.4779 -1.2285 -1.6520 -0.8608  0.2030  0.1577  0.2017 -0.4310  2.2901
#> -0.5963  1.1859  0.4602  0.7288  0.7733 -0.6109  0.0032  0.9620 -0.2274  1.4795
#> -0.6590  0.8292 -0.6882  0.9791  0.2426 -0.1516  0.1243 -0.9430  0.1774  1.2198
#> -0.2925 -1.7998  0.4926 -1.4010 -1.0528  1.0910  0.0685  0.7234  0.9887 -0.3967
#> -0.0577 -0.3962  0.9857 -0.1130 -0.4372 -0.9224 -1.6304 -0.3094 -0.5475  0.4956
#>  0.9176 -0.3547  0.0471 -1.4676 -1.1226  0.5245 -0.2321 -1.4356  1.5577 -1.6457
#>  0.2115  0.1551  0.4133 -1.1233  0.6349  1.1632 -1.2460 -0.7515  0.1219 -0.3514
#> -0.7690 -0.3057 -0.0649 -1.6883 -0.6674 -1.1264  0.4639 -1.7243 -1.6227  1.2493
#> -0.5578 -0.7381  0.3139  0.1430 -0.0167  0.6486  0.5911  0.5378 -0.0365 -0.7085
#>  0.2706 -0.6243 -0.4218  0.1744  1.5342 -0.6664 -0.0911  0.2723 -0.4952 -1.6535
#>  1.9286  0.1036  0.6110 -0.1340 -0.2163  1.6225 -0.0701 -0.4362  1.5496 -1.2914
#>  0.6730  0.0061 -1.0377  0.9165 -1.2679 -1.4774 -0.5682 -0.6067 -1.5910 -0.8733
#>  0.6411 -0.2659  0.3412 -1.2338 -1.5716 -0.1190 -0.5916  0.7768  1.3910 -0.5028
#> -0.3484  0.0129  0.0380  0.8709 -0.4443 -0.6656  0.0235 -0.7319  1.8793  1.0409
#>  0.7036  1.0738  0.0484 -1.6650 -0.1605  0.2006 -0.1744  0.7178  0.2090  1.5741
#> -0.0066 -0.7105  0.5686  0.6669  0.3483  0.4283  2.8652 -0.9253 -0.5278  0.2914
#> -0.1993  0.4763  1.5945  0.8394  0.7594  0.5117 -0.3741 -1.8443 -0.3631  0.3322
#> -0.3417  0.9836 -1.6707 -0.4342 -0.1260 -0.2988  0.7720 -0.3863  1.3028 -0.8621
#> -0.8955  0.7981 -0.7421  0.9693 -1.7067  1.0232 -1.4404  1.3434  0.6120 -0.3598
#> -0.5481 -0.5388 -0.3588 -0.2109 -0.5196 -1.2476  1.5610  0.0429 -0.7635 -0.4673
#> -0.7737 -2.6349 -0.8415  0.0922  0.7391 -0.3676 -0.3602  0.5228  0.3789 -0.1845
#> -0.3638 -0.3322  0.4819 -0.3660 -0.3856  1.4328  0.8103 -0.6687 -0.0243 -1.9178
#> -0.4024  0.8246 -1.0876  0.4823 -0.7530  1.1162  0.0736 -0.1449 -0.0631  0.7479
#> -0.4712 -0.0047  0.0509 -1.6614 -1.3183 -0.5461 -0.5033 -0.1535 -0.8455  0.2750
#>  0.7429 -2.1489  0.3044 -0.3214 -1.8591 -0.0625  0.7753 -0.4002 -0.9893 -1.4467
#> -0.0433  0.7972 -0.1010  0.7682 -0.1640 -0.9391  0.3281  0.8664  2.2910  0.4034
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]