Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 6.5823e-01 -2.0694e+00  2.9249e+00  1.7327e-01  5.6579e-01  4.4443e-01
#> -6.5713e-01  4.4168e-01  5.2315e-01  7.7523e-01  1.8640e-01  9.5630e-01
#>  4.7479e-01  1.4260e-01 -1.5093e+00 -4.8870e-01 -2.4513e-01 -8.3940e-01
#> -7.3329e-01  6.3622e-01  7.9290e-01 -2.3558e+00 -1.5247e-01 -6.4387e-01
#>  1.1469e+00  1.7358e+00 -9.8909e-02  3.1369e-01  9.2112e-02 -1.0986e+00
#>  3.6660e-01 -1.4032e+00  6.5528e-01 -4.6638e-01  4.9129e-01  3.4523e-01
#> -4.4066e-01 -1.2106e-01  7.6896e-01 -2.7117e-02 -2.2389e+00 -1.8945e-01
#> -7.2343e-01  9.3109e-01  4.4010e-01  7.7651e-01  7.8646e-01  7.0823e-01
#> -1.4050e+00  1.0840e+00 -6.0243e-02 -1.3356e-01  1.2472e+00  2.0940e+00
#> -1.1822e+00 -1.9119e+00 -1.6456e+00 -4.9759e-01  1.6797e-01 -2.3830e-01
#> -1.2745e-01 -8.5979e-01 -2.1159e+00 -4.7730e-01 -1.4900e+00 -4.9805e-01
#> -2.6603e+00  7.6268e-01 -3.0831e-01 -2.0608e+00  6.1767e-01 -3.7974e-02
#>  5.3997e-01 -1.0083e-02  1.4324e+00  1.2327e+00  6.3907e-02 -9.1622e-01
#> -3.6996e-01 -6.0256e-01  1.3252e-01 -2.2657e-01 -2.6044e-01 -2.1276e+00
#> -2.3634e+00 -1.1141e+00 -6.6187e-01  9.2739e-01  6.0197e-01 -7.4616e-01
#>  6.1427e-01 -2.3663e-01  1.5682e-01 -4.8215e-01  4.9867e-02 -4.1269e-01
#>  6.5437e-01  3.6187e-02  1.0298e+00  4.1017e-01 -1.7035e-01  6.0378e-01
#>  1.3508e+00  6.5000e-01 -6.7229e-01 -5.5168e-01 -2.1978e-01  7.6631e-01
#>  4.1809e-01  1.1066e+00 -1.4727e+00  1.1629e+00 -2.1330e+00  3.0237e-01
#> -1.2742e+00  3.9310e-02  5.3442e-01  7.0086e-01 -9.3390e-01 -7.8231e-01
#>  1.5386e+00 -1.2282e+00  8.2034e-01  4.4067e-01 -6.6180e-02  1.8307e-01
#>  1.4368e+00 -3.2212e-02  4.5718e-01 -7.7478e-01 -9.3740e-01 -7.7255e-01
#> -1.0636e+00  1.6250e+00  1.1023e-01 -2.3044e-01  3.1100e-01  2.1337e+00
#> -1.0487e+00 -1.2442e+00  7.5378e-01  7.3161e-01  1.8850e+00  3.4651e-01
#> -8.0526e-01 -1.3293e+00 -1.9436e-01  1.1269e+00 -4.4871e-01  2.2208e-01
#> -2.2153e-01 -1.5378e+00  1.2068e+00 -1.0367e+00  1.2680e+00  2.0496e+00
#> -9.9248e-01 -6.2654e-01  4.9230e-01  7.7505e-01  3.5085e-01 -7.4615e-01
#>  1.1681e+00 -1.3163e+00  8.9575e-01  7.0411e-03 -5.4175e-02 -1.8857e+00
#> -7.3877e-01  2.8277e-01 -9.7040e-02  3.3177e-01 -2.8431e+00 -1.6652e-01
#>  1.2943e+00 -9.2169e-01 -2.0830e-01 -5.9717e-02  6.0861e-01 -1.6051e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]