Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-1.6736  0.0976 -0.0584 -0.4558  0.1591  0.3586 -0.5271 -0.5661  0.0537 -0.2383
#>  1.0699  1.0162 -0.2194 -0.0593  1.0322 -0.5038 -0.0686 -1.1713 -1.1524  0.2716
#>  0.2235  0.6576 -0.0018 -0.2256 -0.7648 -0.4394 -0.2332  1.2207  0.2149  0.5801
#>  0.3913  1.4019  1.1406 -0.4554  0.1044  0.9125 -0.8159 -1.3355 -0.7988  0.6861
#> -0.1500  0.3481  0.1755 -1.0197  0.8595 -0.7605  0.2836 -0.2761 -0.8949  1.6685
#>  0.2943  0.2844  1.2011 -0.4272  0.3112 -0.2309  0.1779 -0.0609 -1.3797 -0.5061
#>  0.5889 -0.5336 -0.7225  0.0596 -1.8419 -0.4979 -0.6967  0.5377 -1.2305  0.5022
#>  0.5419 -0.0150  1.7828 -0.2388 -1.0641  0.0033  0.2816  0.1625  0.5214 -1.0033
#>  2.1272 -1.4011 -0.0562 -1.1812 -0.6896  0.6234 -0.0478  0.3917  0.0533 -0.2890
#> -1.6565  1.1351 -0.0363 -0.4976 -0.3828 -0.2625  1.1457  0.3233 -0.2535 -1.3394
#> -0.6473  0.1067 -0.1322 -0.3206 -0.5539  0.2397 -0.7137 -0.3279  0.3891  1.6059
#> -0.2846  0.9066  2.1988 -0.0437  1.6931  2.0678 -0.5094 -0.3871 -0.1914  0.2613
#> -0.8505 -1.0734  1.0184  0.1322  0.4673 -0.6611 -0.7043 -0.1427  0.4390  0.5684
#>  0.2444  0.6823 -0.9153 -0.7953  0.7935  1.9799  0.6026  1.1656  0.7834  1.6707
#> -0.7213 -0.3081  1.1900  0.7592  1.0525  0.3119  0.6521  0.8625  0.3436 -0.5892
#>  0.2046 -1.3276  0.4711  1.8110 -0.4066  0.3541  0.6577 -0.2984  1.1286  0.3180
#>  1.5967 -0.6096 -0.1386  0.0359  1.5899 -1.2430 -0.6747 -0.5769  1.2005  0.4925
#> -0.1294  0.6988 -0.3822 -0.2917 -0.1576  0.5142  0.0297 -0.2550 -2.4990  1.4196
#>  0.5665  0.5917  0.5791 -0.7504 -0.2036  0.9782 -0.5350  1.1601 -0.9271  0.1196
#> -1.5449 -1.2083 -1.2521 -1.8191  1.0962 -1.1498 -1.2178 -0.0771 -1.2509 -0.9565
#> -1.0132 -0.3749 -0.0288  0.6401  1.1020 -1.1459 -0.5009 -0.1381 -1.5892  0.6399
#>  2.3709  0.4147 -0.1109 -1.3685  0.9586 -0.3085  0.9519  0.2048 -0.9986  0.9797
#>  0.1860  0.8093  0.6612 -1.7771  0.2318 -0.3655  0.3368 -0.2174 -0.6337 -0.6113
#> -1.0508 -2.5941 -1.0384 -0.1396  0.6341 -1.8725 -1.4230  0.5570  0.6617  0.5267
#>  1.3702  1.2999 -0.7841 -0.4604  1.6142 -1.4340  0.3251  0.5836 -0.1044 -1.7714
#>  0.1072  1.3595 -0.9087  0.6812 -1.2210 -0.2143 -0.7434 -0.0439  1.3043 -0.4307
#>  0.3946  1.7055 -0.7946  0.5246  0.0287  0.4560 -0.1336 -0.7559 -0.0976 -0.3166
#> -1.5820  2.2407  0.3751  1.3960  0.0896 -0.1695 -1.3421  0.0301  0.1709 -0.9897
#>  0.4103 -1.1255 -1.6872 -0.1315  0.9197  0.2702  0.6877  0.8383  0.2962 -0.7104
#> -0.3539 -0.6651  0.5121 -0.1226 -0.4699  0.3537  0.9264  1.0824 -0.2764 -0.2760
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]