Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-0.5765 -0.6396  1.0172 -0.9133 -1.2912  0.2007 -0.1017  1.3483  1.8162 -0.1670
#>  0.0574 -0.1278 -1.2947 -0.3569 -2.4848 -1.1319 -0.1940  0.2569  0.9609 -0.0794
#> -0.9574  1.9829 -0.6363  1.2151  1.5915  0.9962  1.2626  0.6797 -0.3670 -0.4056
#> -0.4517 -0.5297  0.1967  1.4057  0.1985 -1.6845 -0.1817 -0.8811  0.5948 -0.1623
#>  0.0867 -0.5579 -1.4860 -1.6338  1.3181  0.4130  0.8907 -1.2179  0.6190 -0.1514
#> -0.2209 -0.6810  0.3456  1.4090  1.7110 -0.9218 -1.7543 -0.6033 -0.2326 -0.5280
#>  1.1562 -0.4263  1.7895 -1.2801  1.4380 -0.5708 -0.0300 -0.6897 -0.4854  1.2865
#> -0.6001 -1.3981  0.0686  0.5273  0.0305 -1.1608  0.7169  0.4301  0.5062 -0.5495
#>  2.6695 -0.6280  0.3110 -0.8922 -1.4772  1.3427 -0.6087 -0.0942  0.0937 -1.6861
#> -1.1602 -2.0871 -0.2022  1.5016  0.2314  0.8167  0.0381 -1.0355  1.0106  1.0013
#> -0.2453 -0.2499 -0.7429  0.9886  0.0492  1.7685 -0.1197  0.2573  0.9057  0.2952
#>  0.0836 -0.0409 -0.5737 -0.6176  0.5220 -1.0653 -0.7381  0.2143 -0.7493  0.0399
#>  1.0749  0.7562  1.8316 -1.4634 -0.2718  1.6140  3.2989 -0.1304 -2.2506  0.8889
#>  1.8865 -1.3541  0.5312 -0.2877  0.3840  0.7805  0.5707  1.2666 -0.5140 -1.0144
#> -0.3457 -2.6393 -1.1021 -2.6720  0.4173  1.1857 -1.0746 -0.5698 -1.2080  1.6823
#>  0.2448 -2.4329  2.7132 -0.0669  0.0345  0.6770 -0.3756 -0.0056 -0.7219  1.3990
#>  0.1830 -0.8728  0.9906  0.6317  0.5932  0.5268  1.0896 -0.1713 -0.1287  1.7337
#> -0.0741 -0.5558 -0.9746  0.7933  0.9374 -0.2943  0.2995 -0.3325 -1.6500  1.5126
#>  0.2020 -0.1162 -0.1927  0.1469  0.9030 -0.4253  0.3653  0.3749  0.8639  0.7435
#> -0.0101 -0.2374 -1.2037 -1.2971 -0.0185  0.7404 -1.9919 -0.8745 -0.1715  1.8127
#> -1.1058 -0.2316  1.0358  1.0176  0.8861  1.3438  0.2842  0.2600 -0.3776 -0.8483
#>  1.2828 -0.2462 -0.2197 -0.6064 -0.6628 -1.3676  0.8025  0.8922  0.7296 -1.0225
#>  2.7435  0.5320 -0.0583 -0.0721  0.1178  0.3824  0.2997 -0.6145 -0.0383 -0.0006
#>  1.4430 -1.9804  0.6706 -1.0705  0.1335 -1.0130 -0.5166  0.1373  0.4397  1.4994
#> -1.0219  1.4890  0.4492  0.5519 -0.5906  0.4899 -2.4775 -0.3646  1.0989 -0.7993
#> -0.6261 -1.0538 -0.2656 -1.6192  1.6428 -0.4228 -1.0848  0.1652 -0.4001 -0.3143
#> -0.6700  0.3012 -0.6721 -1.2267 -2.0772  0.3718  0.9023  0.7428 -0.4939  0.9076
#> -1.7461 -1.3520 -1.3137  1.6937 -1.3407 -0.0926 -0.7903 -2.8267  0.1749  0.6607
#> -0.5427  0.6832 -0.0457 -1.3188  0.2787  1.0301  1.6874  0.7507  0.7611 -0.5523
#>  0.0894  1.5424  0.2057 -1.4563 -0.0743 -0.5516 -1.9354 -0.3372 -0.9344 -0.4172
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]