Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.0138  0.6363 -0.0462 -1.1217  0.7008  0.2596  1.1099  0.1272  1.4296  0.0548
#>  0.5410  0.5770  0.5494  1.8241  0.2398 -1.0284 -0.7760  0.2910  1.2876 -0.8416
#> -0.7368  0.3065 -0.6349 -1.2982  1.4570 -1.8707  1.2954 -0.4255  0.3173 -0.2951
#>  0.5177  0.2670 -0.1678 -1.0240  0.2851  1.0204  1.5398  1.9135  1.4785  0.9512
#>  0.2135 -1.1209 -0.7735  0.8201  0.6676  0.6678  1.8674 -1.1190 -0.5997  0.0710
#> -0.8733  0.3423 -1.0645  1.8265 -1.6604  0.2449 -1.2517 -0.8044  1.7520  0.0362
#>  1.0071 -0.3335  0.0705 -0.1180  0.7905  0.4032 -0.5023  0.6840  2.4058 -0.1318
#> -0.7414 -0.7076  0.8690  1.1454 -1.3540 -0.2304  0.3435  1.8130 -0.3893 -1.5423
#> -0.0415  0.0127  0.5790  0.0435  0.7457  0.4512 -0.0067 -0.5307 -1.5055 -0.2243
#> -1.1872  0.9547  1.2033 -0.9445 -0.9624  0.4457  1.6214 -0.5474  0.7248  0.9360
#> -1.9012 -0.5128  0.5532  1.1661  0.2181 -0.7807 -0.1812 -1.1265 -1.9744 -0.6714
#> -0.5043 -0.4589  0.8981  1.6316 -1.6573  0.4137 -0.2076  0.0624  1.8123  1.7847
#> -0.1050  0.2539 -1.6269  0.6201 -0.2452  1.1990 -0.9891 -0.1749 -0.1963 -1.2344
#>  2.6178 -1.2214  1.0232  1.5693  1.1688  0.4326  0.5602 -0.9200 -1.6474  0.6995
#> -0.1527 -1.7237  0.6741 -0.9265  1.1655 -0.4271 -0.5593 -0.9604  0.9399 -0.5875
#> -2.5083 -0.5749  1.7205  1.6205  0.4254 -1.3112 -1.1809  1.0428  0.9702  0.1480
#> -0.3780 -0.6113 -0.3211  1.7592 -0.9096  0.6811  0.6010  0.1939  0.9716 -0.7635
#> -1.7522 -0.0953 -1.3144 -1.2773 -0.1713 -0.3524 -0.7449  0.9236  1.2799  1.7400
#>  1.3205 -0.3779 -0.6547  0.0627 -0.4896 -0.5019 -0.8565 -0.1463 -1.9036 -0.7833
#>  1.6578  1.2836  0.0161  1.2112 -0.4679  0.0698 -0.4816  0.3072 -0.9742  1.2438
#> -0.7733  1.4163 -1.3304  0.7725 -0.7280  0.4597 -0.8975  0.2379 -0.1823 -0.4689
#> -0.0624  0.4162  0.4693  1.7641 -1.8361  2.7587  0.6000 -1.2522 -0.9399 -1.4477
#>  0.9202 -0.2476  2.5076 -0.0193  0.7960  1.3859  1.0655  0.0551  1.1840 -1.4858
#> -0.8160  0.3384 -0.0430  0.3599 -1.1521 -2.1861  0.8497 -0.4002 -0.2731  0.1805
#>  0.5640  1.9766 -0.8569  1.3160 -2.2848  0.3347  0.3859  0.0307  1.1109  0.6970
#> -0.5731  0.9173  1.5766  0.6717 -1.3758 -1.6178  1.6624  1.2846  0.5139 -0.1387
#>  1.4165 -0.2935  0.1680 -0.4398  0.1190 -1.0184 -0.1399 -0.7096 -1.0376 -0.4626
#>  0.8992 -1.2095 -2.6518 -1.1385 -1.8290 -1.5624  1.9038 -1.7087 -1.0560 -0.1325
#>  1.4133  1.2904 -0.7597  1.7095 -1.5051 -0.6702 -0.8195  0.8855  0.6110 -0.6622
#> -0.0387  0.2078  0.2156 -1.3005 -0.0821  0.3267  0.3942 -0.1553 -1.1209  1.1593
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]