Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.1975  0.1363 -0.5702 -0.0151  1.0552  0.1658  0.4409  0.0485 -0.3640 -1.5562
#> -0.9644  0.9965  0.6831  1.0654  0.2200  1.0254  0.5855 -0.7416 -1.3989  0.7877
#>  0.8300  0.9539  0.2980 -0.8992 -0.4128 -0.2817 -0.6700 -1.0377  1.2535 -0.8246
#> -0.4476  0.3860 -0.3088  1.1196  0.4673 -1.2704  0.1876  1.1804 -1.5134 -0.3262
#>  0.4924 -0.9562 -0.7458  0.4491  0.6308  1.2891 -1.7293 -1.2599  2.1107 -0.2023
#> -1.7858  0.3321  0.2279 -2.9762  0.3026  1.7035  0.2543  1.5243  1.2796 -0.3059
#>  0.0475 -1.0091 -1.2365  0.6934  0.9840 -1.0649 -0.1151  1.0893  0.1180 -0.1804
#>  0.8116  0.6188 -1.7343 -0.1277 -0.4545  0.8333  0.2771 -0.8468 -1.2034  0.8498
#>  0.5610  1.3646 -1.7347  1.4425 -1.1005  1.2000  0.2502  1.1043 -0.1362 -1.1972
#>  0.3774  0.4790  0.5940  1.1127  0.2305  0.0280 -2.7400  0.5515 -0.2985 -0.5540
#>  1.7556  1.2312 -1.1898 -0.8038  0.8663 -1.9170  1.9513  0.8944  0.1931 -0.5963
#>  0.4628 -1.0915 -0.4243  0.3260  0.0133  0.3061 -0.4002 -0.3110  0.5482 -0.7325
#>  1.0581 -1.1467  0.4726  0.4326 -0.8209 -0.3506  0.2272 -0.9898 -1.2534 -0.2001
#>  1.1539  1.8769  1.1133  0.0740  0.4854 -0.7244 -1.5447  1.1827  0.8404 -0.3763
#> -1.1509 -0.1224 -0.6268  0.0929  2.3843  0.1208 -0.6676  0.6381 -0.2542 -0.7784
#> -0.6489 -1.2957  1.8299  0.1618  1.2342  0.0168 -1.0442 -0.1499  0.2994 -1.9764
#> -0.9372 -0.1464  0.0195 -1.2885 -1.5547  0.3102  1.3669  0.9203 -0.7020 -0.0838
#>  0.0019 -1.9856 -1.4276  1.0406 -0.5169 -0.0178  1.4171 -0.0836  1.4172  2.5724
#> -0.7518 -1.2029 -0.9766 -0.0478 -0.8344 -0.9468 -0.4773 -1.1574 -1.4516 -1.0318
#>  1.9486  0.2762 -1.2751 -0.2695 -0.3225 -0.4082  0.3060  0.5094 -0.1473  2.3010
#>  0.8208  0.5015 -1.2949  1.4148 -0.2067 -0.0233  1.0697  1.2085 -0.0349 -0.4188
#>  1.3352  0.3111 -0.4636  0.0022 -0.1868  1.2533 -0.9030  0.6282  0.1818  2.0706
#> -0.6466 -0.0850  1.2073  0.3469 -1.4312  0.8662  0.5445  0.1222 -0.5739  0.4042
#> -1.0077 -0.6226 -0.5374  2.0082 -0.4040  0.5349 -0.5432  1.2952  0.5897  0.5398
#> -1.3747  0.7969  3.7019 -0.9090  0.1866 -1.3502  0.5442  0.3327 -0.9113  1.6786
#>  0.1924 -0.2084 -0.9977  0.7405 -0.3283 -2.6870  0.4346  0.5608 -0.0403  0.4078
#>  0.3480  1.0117 -0.0373 -1.2577 -0.9827 -0.4808  0.6124 -1.2771  0.5928  0.0152
#> -0.9271 -0.3477 -1.4638  0.4029 -2.0729 -0.6151 -1.1032 -1.2645 -0.1845 -1.7900
#> -1.8946  1.4804 -0.5535 -1.8392 -0.1275 -0.2850  0.8568  0.0706  0.1000  1.2136
#>  0.2152 -1.6815  0.2982  1.1826 -2.1365  1.0074 -1.2775  1.0697 -1.0114  1.2504
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]