Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.8961  1.3125 -1.1855 -0.6450 -0.3888  0.0993  0.0915  0.3411  0.2227 -0.8675
#>  0.0482 -0.3647  0.6932 -0.6964  0.0606  0.8512 -2.2046  0.3840  0.2374 -1.8325
#> -0.0727 -1.4566  0.7223  1.1038  1.0064 -0.3838  0.0203 -0.1987 -1.0991  0.6554
#>  0.2473 -0.8229 -0.0922 -0.3655  0.7656  0.4108 -0.5222 -0.1409 -1.1797  0.1995
#> -0.6406 -0.2438 -1.0156  1.3686 -0.4113  0.6942  0.0619  0.2149 -0.4306 -1.5514
#> -0.6934 -0.0547 -0.1216 -0.7676 -0.1282 -0.2701  0.1595 -0.2884 -0.4173  0.0788
#> -1.4146  1.0236 -0.0703  1.9326  0.8623 -0.7921 -1.1602  0.4626 -0.3609  1.3669
#>  0.4612  1.0459  0.0308  0.6508  0.9504 -0.1823  0.6262  0.7697  0.6183 -1.3439
#> -1.4803 -0.7457 -0.6544  0.0394  0.7310  1.1534 -1.8515  0.1218  0.8467  0.5940
#> -1.1973 -0.2574  0.2035 -1.2305  0.2572  0.5159 -1.1651  0.4165 -2.0750  2.2121
#> -1.7634 -0.4952  1.6065 -0.3230  0.0614  0.4314  0.4754  0.4061  0.1112  0.7559
#>  0.5010  0.4191 -0.5069  1.0304  0.1571  0.9657 -0.7015 -0.6073 -0.8913  1.2866
#> -0.2131 -1.1248 -1.3635 -0.3157  2.4571 -0.3797 -0.5882 -0.7077 -0.4533 -0.1520
#> -0.0196  0.9982 -1.1238 -1.2421 -0.9183  1.9902 -0.0395 -1.9735 -1.2588  1.3138
#>  0.7659  0.2422  0.2021  1.1042 -0.9601 -0.6425 -1.7044  0.3113 -0.0796  0.8676
#> -1.7849 -1.5461 -1.7550  0.5795 -1.1382 -1.2025 -0.2725  0.0663 -0.6481  0.8164
#>  0.5064  2.1330 -0.5477  0.2947  0.8880 -0.0528  0.1810  0.0268 -0.7809 -0.8240
#> -0.3625 -0.4165 -1.8687 -1.5184 -0.3861 -1.2728 -0.7179  0.2753 -1.3846  0.5716
#> -0.6489  0.1198 -0.4414  0.6814  0.5486  1.5345 -0.9595 -0.4554 -0.0019 -1.0855
#>  2.7599  0.5692  0.3694 -0.4370 -0.6528 -0.6870  0.0480 -0.3219 -0.1756 -1.3709
#>  0.0609 -0.8376 -0.5744 -0.7348 -0.7568  0.8103 -0.9877  0.9746 -0.3725  0.8124
#> -0.1213  1.4531 -0.4929  0.4165  1.5614 -1.5473 -0.1639 -0.4509  0.1719  1.2943
#>  0.6707 -0.5402 -0.1549  0.5650 -0.6553 -0.1484 -0.7786 -1.6523  0.4715 -0.5571
#>  0.6554 -1.3804  1.9208  0.9156  0.1422 -1.3381 -1.0916  0.3071  0.0187  1.7929
#>  1.0773 -0.1002  0.4767 -1.1004  1.0670  0.4232 -0.7582 -0.7061  0.9749  1.3096
#>  2.3678  0.2533  0.3377  0.2611 -1.3816 -0.0855  0.0496 -0.8361 -1.3402  1.1886
#>  1.5331  1.4031  0.1401 -0.3006  0.1586  2.8832 -0.7083  0.8000  0.8549  0.9131
#>  0.4073  0.3357  0.7061 -0.1620 -2.0491  2.2503  0.2950 -1.6415 -0.0736 -1.6895
#>  0.7403  1.7063  0.3053 -0.6139  0.3093 -0.1669 -0.9033  0.7577  1.9133  1.5058
#> -1.7284 -0.1337 -0.4431  0.3541  0.7347  0.6887  2.2706  0.8055  0.2164 -0.5867
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]