Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.2834  0.0519 -1.0840 -0.4660  1.2438 -0.4125 -2.2425  1.4000  1.6101  1.7624
#> -1.9654 -1.4588 -1.2030 -1.3124  0.1463  0.2007  0.9883  0.7516  1.0895 -0.7467
#> -0.4941 -0.5118  1.2944  0.7692 -0.4973  1.0781 -0.8611  0.9996 -0.5614 -0.7037
#> -0.2316  0.0795 -0.6012 -0.9953 -0.5653  0.9587  1.4715  1.4456  0.3971  0.9960
#> -0.6215  1.2067  1.0922 -0.4532  0.9893 -1.5648 -0.0911 -1.2118 -1.0886  0.7682
#> -2.1193  0.7935  0.6272 -0.6307 -0.5717 -1.2534  1.4168  0.7387  0.5244  2.5145
#>  1.3855  1.1218 -0.9812 -2.7794  0.3207  1.2952  1.1444 -0.5104 -0.7603 -1.1336
#> -0.5076  0.0822 -1.2110  0.1659 -1.1551  0.4827 -1.8277 -0.0295 -0.5959  1.1296
#> -0.0973  0.9184 -0.4696 -0.1460  0.1027  0.1904 -0.6890 -1.3047 -0.2977 -0.1142
#> -0.1082  0.3769 -0.7455  0.2252  0.2548 -0.8385 -0.3356 -0.7524  0.0789 -1.7324
#> -0.1709  1.2562  1.6303  1.5818 -0.3153 -0.5659  0.6751 -0.6338 -0.3816  0.4199
#>  2.0517 -0.3436 -0.9492  1.1734 -1.0126  0.9734 -2.4357  0.8823  2.7234 -0.2763
#>  1.0848  0.9682  0.0752  1.9238 -0.3956  1.0371 -0.0176  0.0530 -1.2685  0.2009
#> -1.0142 -0.3812 -0.6412 -0.1283 -1.4435 -0.5677 -0.5030 -3.1669  0.7460 -1.5092
#> -0.0301 -1.5790 -0.6311  0.1657  1.0584  0.0737  0.4975 -0.5673 -1.5469 -0.8983
#>  0.3088  0.5466  1.0251  0.5783 -1.6141  0.0103 -0.1549 -0.2052 -0.6438 -0.7679
#>  1.5416  1.5100  0.7828  1.0884  1.4462 -0.0181 -1.4780 -0.8455 -0.7138 -0.0806
#> -2.1537 -0.0968 -0.7106 -1.3505  0.2610 -1.4850 -1.4820  0.4458  0.8279 -2.5574
#> -1.3470  1.6128 -0.0472  0.7917 -1.7557  0.6555  2.7213 -0.0187 -1.0007 -0.4555
#> -0.3585 -0.3114  1.3694  1.0532  0.4573  2.2333 -0.5299  0.0174  0.8500 -0.4613
#>  0.1590 -1.4860 -0.7412 -0.8566  0.2562 -0.4007  0.0640 -0.1551  0.3253 -0.7590
#> -1.5600 -0.5588 -0.6740 -0.0751  0.7784 -0.0705  0.6136  0.1043  0.4571 -1.5254
#>  1.2598  0.5892  1.0764  0.8806  0.4587  0.9171 -0.2775 -0.4509 -0.3376  0.4237
#> -0.7023  0.8740 -1.2045  0.7040 -1.9007  0.8286  0.2466 -0.4486  1.2716  0.9879
#>  3.0323 -1.0999  0.2482  1.2679 -0.1805 -0.9478  0.7078  1.5048  1.1038 -0.5725
#>  0.4909 -0.9662 -0.9461  0.9664 -1.6204 -0.1724  0.9342 -0.0965 -1.0472 -1.2322
#> -0.0729  0.9080  0.5497  0.2030 -1.1166  0.1429  0.3219 -0.1429 -0.1047  0.1220
#>  0.1384 -0.9110  2.8819 -0.9899  1.4824 -0.2438 -0.5241  1.3945  1.2450 -0.1991
#> -0.4867  0.5369  0.4223 -1.4958  0.3102  0.2780 -0.4311  1.9263 -0.7177 -1.4488
#> -0.2603 -0.1559  2.3844  2.8795 -0.2135 -0.7226  1.9380  0.7842 -1.0008 -0.1473
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]