Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0518 -0.2879 -1.7103 -1.2761  0.2988  1.2769  0.5602 -0.3063 -1.6725 -1.0096
#>  0.0686 -0.0312 -0.0208  0.2671 -0.0616  0.4808 -1.2544  1.9039  1.5164 -0.9015
#> -0.4464 -0.3176  2.2035 -2.4737  0.0748  0.3414  1.0242  1.7764  0.7062 -0.0132
#>  0.7187 -1.0076  0.0128 -0.4883  1.0464 -0.0794 -0.0978  0.1247  0.3921 -0.9118
#>  0.7533 -1.8579  0.7657  2.3277  0.4648  1.6589  0.6021 -1.7128  0.7469 -0.8006
#>  0.1180 -0.2561  0.2854 -0.1504 -0.2841  0.7878  0.5099 -0.9000  1.3692 -0.0144
#> -0.6214 -0.8327  0.3597  1.5874 -0.3166 -1.6237 -0.0457  1.3961  0.0268 -0.0867
#>  0.0146 -0.1041 -1.4005  0.2894  0.4907 -1.9061  0.0552 -1.3181  0.4194 -0.3745
#>  0.1045 -0.1752  1.1015 -0.9364 -0.2335  1.9518 -0.0049  0.2910 -0.0612  1.0870
#>  0.6485  0.2294 -0.4581 -0.1712  0.9002 -0.7104 -0.1567  1.7536 -0.3000  1.3799
#>  0.2610  0.4463 -1.1303 -0.2345  1.4595  1.7515 -0.7041 -1.2767 -0.1823 -1.6989
#>  0.2188 -1.0053 -0.8758 -1.2751  1.8234  0.7883  0.6889  0.7972  1.6670  1.1819
#>  0.1601 -0.1881 -0.7998  0.0582  0.7979  0.2309 -1.5910 -0.6351 -1.8183 -0.4555
#> -0.0219 -0.3997  1.7758 -1.5038  0.0971  0.2553 -0.6292 -0.8906 -0.0535  0.1639
#> -0.1027  0.2771  0.3270 -1.2041 -0.4850  0.6069  1.1983 -0.2429 -0.1475 -1.3936
#>  2.6068  0.2918  1.5908  0.1986 -0.3171  1.1844 -1.5918 -0.2659  0.9860  0.4465
#>  0.6457 -0.1727 -0.0549  1.2706 -1.2638 -0.9823 -0.1236  0.2498 -0.0147  1.2321
#> -0.6269 -0.2976 -0.5248 -0.2368  0.8509 -0.9118 -0.7279  0.5475 -2.4331  0.7582
#>  1.9431 -0.6466 -1.7617  0.1927  1.7258 -0.0333 -1.2712 -0.6798  1.0756 -0.0077
#> -0.4317  0.7423  1.1630  0.3174  0.2052  1.1892 -0.7598  2.5291  0.1101 -0.0186
#> -1.1320  0.1521 -1.4070  0.5822  1.0107  2.7152  1.3545  0.5031 -1.0193  0.4200
#> -1.7775  1.5303  1.1077  0.8178  1.9110  0.5547 -1.0839  0.3545 -0.0547 -1.2503
#> -0.2023 -0.0012  0.4628 -0.3795  0.3042  0.4084 -0.4762  0.0487 -0.1916 -1.1142
#>  1.2170 -0.2363  0.8907 -0.0232  0.8366 -1.5512 -0.2858 -2.2394 -0.0865  0.1636
#>  0.9425 -2.2743 -1.0313 -1.1521 -1.3823  1.7820  0.0456 -0.4674 -0.4572 -0.5481
#> -0.5676 -2.2377 -0.6009 -1.1123  0.4257 -0.7217  0.3173 -0.3111 -0.8425  2.0910
#> -1.0604 -2.5904 -1.0784 -1.2254 -1.2822  1.2141 -0.8709  1.9006  0.3220 -0.6278
#>  0.2265 -0.4948 -0.2964 -0.2604  0.8540  1.3505 -0.1914  0.8156 -0.9952 -0.8046
#> -0.4498  1.7362 -0.7996 -0.8102 -0.3411  0.3395  0.6693 -0.7677  0.5676  0.3104
#>  0.6251  1.2495 -0.4050  0.6689 -0.3621 -0.6474  0.6886  0.6253  0.4025  1.1837
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]