Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7567 -0.5988 -0.1386 -1.1623 -0.2147  1.3030  0.0823 -0.1790  1.4975 -0.3962
#>  1.0721  1.6521 -0.0912 -0.1511  1.6095  0.0761  0.2536 -0.3202 -0.1235 -0.1725
#>  0.6349  1.6455  0.3823  1.0589  1.5274 -0.2384  0.4417  0.0269 -0.7888 -0.9275
#> -0.0463  1.5305  0.4013  0.1596  0.5246 -0.9620  0.3830  0.0752  0.8636  2.0816
#> -0.3526 -0.2296  0.4240 -0.3024 -0.1462  0.0119 -0.4934 -0.2543  0.5664  0.2319
#> -1.4273 -0.7056  0.3131  0.2021  0.4535 -0.7314  0.2409  1.8042  1.4112 -1.0558
#> -0.8031  0.5268  0.0638 -0.4140  0.1145 -0.7102 -0.6635  0.4252 -1.1473 -0.7560
#> -2.2452  0.2658 -0.5824  2.4460 -3.0219  1.5613  0.2459 -1.3483 -1.1535 -1.3614
#>  0.1361 -1.1496 -0.2418 -1.3401  1.1974 -0.2698 -0.0787 -0.6813 -0.9623  0.1834
#>  0.1614  0.0504 -1.3519 -1.4879  0.9107 -0.1208 -1.3635 -0.7297  0.0761 -0.7343
#>  0.1991  0.0106 -0.5328 -0.6875  0.5590 -0.7711 -1.0119 -0.3633 -0.8525  0.0187
#>  0.1584  0.3217 -0.6002 -0.0535 -1.3622  0.0340  0.7926 -0.1052  1.2083  1.0924
#> -1.4757 -0.4288 -0.2858  0.6294 -0.7096 -0.4328  0.4349 -0.1776  1.3832  0.7670
#> -1.6920 -0.1855 -1.6463  1.9172 -0.9743 -0.5823  1.7168  0.9505  0.5479  0.7692
#>  1.1354  1.1656 -0.0094  1.1859 -0.7166  1.3567 -0.7808 -0.5306  0.3293 -0.4349
#>  0.0145 -0.3003  1.3213  2.0579 -0.2171 -0.7446  1.3161  2.1582  1.1262 -1.4839
#>  0.0416  0.6326  0.5245 -0.9761 -0.6854 -1.4080 -0.4428  1.0267  0.6720 -0.3255
#>  0.3516 -1.0836  0.0068 -0.0755  0.7463 -0.6241 -0.9887 -0.6371  0.1379  0.1638
#> -0.7305 -0.6154  0.7252  1.6427 -0.0910  0.0925  0.6563  0.0494  0.3041 -0.1300
#> -0.5042 -1.3360  0.3268 -0.4285 -1.3530  0.1900 -0.5748 -0.1345  1.6881  0.9700
#> -0.6465 -0.3949 -1.1104 -0.6223 -0.1590 -0.0856  0.3127 -0.4303  0.4696 -1.0954
#>  0.2354  1.6841 -0.2891  1.8578  0.7954  1.0428 -1.2042  1.2080  1.2208  0.2405
#> -1.7583  0.5833  0.2086 -1.4170  0.2385 -0.3295  0.7086  0.5559  0.0515  0.6454
#>  0.5188  0.5519 -0.8012 -1.4867  0.1952 -0.7053 -0.7856 -0.2799 -0.8646  0.1657
#> -0.7951  0.0174  0.8454 -0.9156  1.8450  0.2265 -0.0105 -0.1249  1.4181 -0.6898
#> -0.8941 -0.7285  1.6071  1.3542 -0.9072  0.4731 -1.2043 -0.4981 -0.9857 -1.7360
#>  0.5904  0.1202 -1.3134  1.3813 -0.1089  0.7078 -1.7637  0.4283 -0.8904 -0.9523
#> -1.1857 -0.8745  0.8681 -0.6736 -2.6399 -1.7877 -0.6535  1.0992 -1.2326 -0.6414
#>  1.2500  0.2904  0.4048 -0.5196  0.8955  1.9274  0.8854  0.2860  2.3217 -1.2369
#>  0.1256 -0.4199 -0.1362  0.6675 -0.4091  0.5913  0.0887  0.0800  1.1342 -0.6980
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]