Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0586 -0.4577 -0.6572  0.5239 -0.6340  0.1088  0.3944 -0.7198 -0.6342 -0.9533
#> -2.5171 -0.5051 -0.3059  1.0633 -0.8199  0.8904  0.5164 -1.9845  0.1739  0.7825
#> -0.2440  1.4431 -0.0598 -0.5698 -0.7867 -0.6994  0.2336 -0.4535 -1.3765 -0.3545
#> -0.0365  0.3369 -1.3859  0.6360 -0.3260  0.0723  0.9773  1.5135  1.7087 -1.2511
#>  1.2487 -0.0150 -0.2080 -0.3279 -0.5864  0.0176  0.8886 -0.8303 -0.0719  0.4943
#>  1.7115 -1.1918 -0.9538  1.7974  1.2186 -0.1305  0.2807  0.5271  0.4059 -0.8003
#> -1.0030  1.2456  0.3334  0.4454  0.3979 -0.4756  1.4849 -0.3878 -1.1364  1.4588
#> -0.4276  1.2074 -1.7308  0.0300 -0.6899  2.0308  0.3496 -0.6293 -0.2347  1.2837
#> -0.7439 -0.6034 -1.9821  0.6285  0.0606  0.1620  0.9801 -0.3189  0.9825  0.3292
#> -1.8190  0.0111  0.9823 -0.2089  0.5257 -0.2130 -0.2213 -1.1152  0.6024  0.1933
#> -1.7286  0.2753 -0.7113 -0.1297 -0.5303 -1.0320  0.5105 -0.3978 -0.3945 -0.1831
#> -0.2872  0.8352 -0.5238  1.2550  0.2145  1.5728  0.3276 -0.4853 -1.2656  0.0855
#>  0.0664  0.7911 -1.0858  2.6015  0.7073  1.3751  2.7366 -1.3323 -0.2982 -0.3868
#>  0.7520  0.6767  0.6857 -0.7773 -0.0592 -0.6341  0.8467 -0.2773  0.6492  0.4444
#> -1.6877  1.4170  0.5582 -0.4820  0.7146 -0.7320 -1.7723  0.7760 -1.6758 -0.0974
#> -0.7030  1.3897  0.8545  0.2108  0.4493 -1.1251  0.7482 -0.9156 -0.5680  0.5738
#> -0.1027 -1.4129 -1.8526 -0.1596 -0.5906 -0.9714  0.2152 -1.4236 -0.3375  0.0165
#>  0.6800 -1.1795  0.7855  0.7433 -0.7955  0.6159  0.2305 -0.8965  1.6027 -0.3642
#>  1.3402 -0.5771 -0.1413  0.6085 -0.6335 -0.3199  1.5012 -0.0474  0.4189 -0.8046
#>  0.0981  0.7954 -0.1223 -0.0987  0.3292  1.0127 -0.8833 -0.7053  0.1429  0.7628
#> -0.4074  0.3521  0.7772 -1.3259  0.0922  0.3375  0.8111  1.3594  1.2787 -3.2147
#> -1.0586 -0.2423 -0.4771  0.9226  0.6938  0.3959  1.6484 -0.6054  0.2988 -0.3370
#>  0.1748  2.4163 -0.5023  0.2400  0.2796  1.2503 -1.8638 -0.3606 -1.4573  0.9960
#> -0.6684  2.1143  0.4699 -0.2425 -0.4939 -0.4897 -0.9551 -0.1026 -0.0174 -1.5902
#> -0.8154  0.5627  0.4998  0.5407 -0.3012 -1.5356 -0.5869  0.4827  0.6297  0.0904
#>  0.1486  1.2517  0.4192  0.2516 -0.6017 -0.3649 -0.4681 -2.1734 -0.7471  0.7251
#> -0.8808  1.0414 -0.8187 -1.9763 -0.0071 -2.0011 -0.8046 -0.2813 -0.2565 -0.0005
#> -3.1878  1.1000  0.7942 -0.3545 -0.5084  0.6480 -1.9096  1.0483 -0.7729 -0.9195
#> -1.9184  1.1028 -0.8140  0.8563 -0.4300  1.3855 -0.9462 -0.4540 -0.7230 -0.8957
#>  1.8057  1.9433 -0.3243 -0.1847  1.1182  0.8467  0.2025 -0.2831  0.3417  0.8635
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]