Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.0693 -0.1518  0.9454 -1.0459 -0.7368 -0.0472 -0.2585  0.3871 -0.6921  1.0517
#>  0.5480 -0.5515 -0.2866  0.1387 -0.0046  0.5381 -1.3995  0.7519 -0.5567  0.2553
#>  0.0888  1.3332 -0.2166 -1.3731 -1.7506  1.0415  0.5209  0.1479 -0.2042 -0.5874
#>  2.0244 -0.3916 -0.5205 -0.7269  1.3065 -0.9858 -0.3237 -0.0770 -1.1608 -1.1313
#>  0.6693 -0.5579  0.6665  1.5187  0.1677 -0.6147 -0.2764 -0.7559 -3.2681  0.7147
#>  0.3728  1.6885  0.0029  0.4712  0.4049 -0.0931  0.3434  0.3993 -0.1869 -1.7701
#>  0.7702  0.4781 -0.9741 -0.3849 -0.4413  0.7158  0.0901 -1.5211  0.2609 -1.0099
#> -0.6484  0.9898  1.2271  0.4717  0.4035  1.0667  0.5749 -1.1534 -0.3824 -0.5203
#> -0.9732 -1.1708  0.1161 -1.7306  2.3656 -0.4926 -0.1655  0.0193 -1.3408  0.6881
#>  0.9261 -1.3931  1.2423  0.5289 -0.5828  0.7500 -0.1051  0.3261  1.2092  0.9480
#>  1.2839  1.3169 -1.4987  0.2240 -0.2355  0.2785  1.6671  1.1722  0.0800  0.9823
#> -0.2728  2.6266  1.3225  1.2825  0.6805  0.3703  0.1547 -1.1943  1.2475  0.9663
#> -0.7696  0.2348 -1.4840  0.5170 -0.5158  0.4675 -0.2005 -1.4208 -0.2638 -0.2711
#>  1.1336 -0.5462 -0.6738  0.3325 -0.0686  0.5541 -0.2075  0.3624 -1.1426 -0.1188
#>  0.7262 -0.8559  0.0652  0.2454 -0.2346  1.8075  0.9806  0.4152  1.1382 -0.3998
#>  1.4084 -0.1163 -0.3476  1.4530  1.8517  1.3870 -0.6641  0.0844 -0.1264 -0.4516
#>  0.7377  1.2372 -0.8876  0.6292 -1.2062  0.8290 -0.1293 -0.0375  0.5030  1.1767
#>  1.5720  1.4782 -0.6854 -0.6803  1.9388  1.3458  1.0468 -0.8556  1.4524 -0.5042
#> -0.7283 -1.8454  0.1241  1.2728  1.4102  2.1257 -0.1062  1.6580 -0.8802 -0.2877
#> -0.7014 -1.1906  2.1979  0.5228 -0.8413 -0.8695  1.9606  0.1621 -1.6142 -0.8656
#>  1.1834 -1.6885  1.2614 -0.2269  1.3485  0.3535 -1.2762 -0.4683  1.2391 -1.9962
#> -0.4967  0.6109 -0.4933  0.9072 -0.5105  0.9848 -0.1378  0.6303 -0.9165 -0.1141
#> -0.5055  0.9019 -0.0804 -1.1037  0.3442  0.7576  0.9446 -0.5600  1.6882  0.1468
#> -0.6130 -0.5159 -1.4472  1.1179 -0.4732  1.3049  0.2679  1.2856 -0.6780 -0.6517
#> -0.4918 -0.7768 -0.6151 -1.2005  0.9550 -1.4457  2.4159  0.2530  1.7712  1.1319
#>  0.2231  0.8775  0.2719  0.4401  1.4144 -1.9500 -1.0983  2.3958  0.5479 -0.7042
#>  0.5692 -1.6299 -1.7195  0.8603 -1.1805  1.8444 -1.4716 -0.1196 -0.7398 -0.2851
#>  1.2963  0.7228 -0.4523 -0.4063  0.4794  0.6270  0.6225  0.4241 -0.4730  0.6283
#>  0.6509  0.4067 -0.8094  1.6056 -1.5041  0.7543 -1.0005  1.1728  0.8713  2.5363
#> -0.7180 -0.7514 -1.8243  0.6694 -1.3081 -0.7784 -0.5786 -1.7585 -0.1773  0.1009
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]