Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4309 -0.6825  0.6434  0.5701  0.4700  0.2565 -1.6332 -0.7498 -1.4133  1.1551
#> -0.3602  0.9492 -0.7439  2.2132 -0.6557 -0.0314 -0.2659 -1.2439  0.8252 -2.4473
#>  1.0856  1.1149  0.0387 -1.2779 -0.7032  1.1333 -0.0897  0.7850 -0.3993  0.1194
#>  0.9838 -0.0002 -1.7926  0.1798 -1.0879 -0.0538 -0.8856  1.2190  1.2472  2.0204
#> -1.1287  1.6714 -0.4512  0.0445 -0.8935 -1.5451  0.0585 -0.3777  0.4397  1.4047
#>  1.3683  1.1044  0.6527  1.6349  1.0011  0.0380  0.5469  0.8247  0.1990  0.4916
#>  0.3008  0.4466 -0.0738 -0.8983  0.5889  0.0251 -0.6937 -0.1740  0.3079 -0.3069
#>  1.2038  1.0961  0.2792 -2.1414 -0.1306 -0.8728  0.6658  0.2086  0.3150  0.4118
#> -0.8859 -0.4569  0.5372 -0.1127 -0.1878  0.3715  1.1295  0.0302  0.8248  1.5506
#> -1.1215 -0.8363  0.4088  2.9495  0.1187 -0.2439 -0.8694  0.0911 -1.0626  0.0296
#> -1.4172  1.6940  0.7154 -2.4892  0.0798  0.6092 -1.8883 -0.0103 -0.9780  0.1147
#>  0.0051  2.2400 -1.2407  0.4988  0.9266 -0.3141  0.2932  0.0429  0.0505 -0.7668
#> -1.6432 -0.9458 -0.4227 -0.2083 -0.8528 -1.0588  1.8375  0.1979  0.4226 -0.4359
#>  0.5939  1.5530  0.1331 -1.4326  1.5571  0.9175 -0.3417 -1.5938 -0.3552 -0.3212
#> -0.8994  0.4053  0.1226 -0.4776 -0.1263  0.3723  0.7334 -0.0189 -0.7011  0.8150
#>  1.4580 -2.2050  0.1880  2.0244 -1.4481 -0.2094 -1.2844 -0.4300 -0.2107  0.5566
#> -0.5707 -1.5401  1.0677  0.9103 -0.0806  0.2690  1.0227  0.3888 -1.1327  0.4259
#>  0.7918 -0.3424 -0.0532  0.7426  0.3145 -0.7836  0.3002 -1.1982 -1.4196  1.1397
#>  0.3516  0.0559 -0.0500  1.2456  2.0667 -0.2223 -1.1527 -0.1604 -0.5435  0.2143
#> -1.6515  0.7281 -2.4244  1.5187 -0.7707 -2.0896  0.4233  0.0620  0.6624 -0.9210
#> -0.2408  0.0503 -0.8384 -0.5131  1.0071  1.1485 -0.4059  0.1251  0.9840  0.1383
#>  0.0163  1.4686 -1.7628  0.6430  0.5872  0.4552 -1.3636  1.2806 -1.2572 -0.2297
#>  0.2366  0.4271 -1.1784 -0.1473  0.3868  0.6572 -0.6241 -1.2327 -1.1616 -0.4940
#>  0.1594 -0.0984  1.8281  0.8056 -1.5566  1.3669  0.4014  0.7303 -1.0758  0.1247
#>  0.1677 -2.4912  0.0297  0.9851 -1.1960 -2.0557 -0.2150  0.1531  0.3095  0.5414
#> -0.4224 -0.2303 -0.9927 -0.5304 -0.8688  0.6644  1.0450  0.4777  1.7047 -1.3462
#>  0.2517 -1.0390 -1.1111  1.7078 -0.3871  0.7816  0.6556  1.3966 -1.6485  1.0068
#>  2.0352  0.9090  1.5877  0.6409 -0.0913 -0.2819 -1.6140 -0.8974 -1.8279  0.1433
#>  0.9235  0.4084  0.6034  0.5688  0.6593 -1.2780  0.0412  0.2658  0.5744 -0.3702
#>  0.3436 -0.5410 -0.4570 -0.2489 -1.3169 -1.0865 -0.6521  1.1655  0.8389  0.7934
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]