Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3555 -0.1142  1.0449  1.6081  0.2051  1.3705 -1.7226  0.4361 -0.4842  0.7591
#> -1.1545  0.1486 -0.7709  0.3604  1.1619 -1.7421  0.6767 -0.1736 -1.8390  1.1875
#>  0.3213  0.6256  1.8438 -0.4999 -0.3689 -0.1936  0.8039  2.4178 -0.3986 -0.1259
#>  0.3702  0.5477  0.2651  0.2745 -1.0157 -0.3483 -0.1384 -0.6607 -2.1890  1.0789
#>  0.1930  0.7483 -0.0059 -0.6594 -1.2692  0.1335 -0.7549 -0.1412 -0.8491 -1.5880
#> -0.0501 -1.5439  1.8523 -1.4825  1.0594  0.2199 -0.8158  0.9169 -1.2238 -0.7618
#> -1.4605  0.5817 -0.1333 -0.7738 -0.9453  0.3655  1.0178  0.5185 -0.5846  0.0925
#>  0.9126 -1.6128  0.8695 -1.2482  0.6771  0.1963  0.8236  1.2707  0.4634 -0.8685
#>  1.0722 -0.4317 -0.0518 -1.1256 -1.6797  1.5070  0.5025  1.2069  1.1906 -2.2011
#> -1.1272 -0.1096  0.3647  0.2075  2.3064 -0.1050 -0.4971  0.5003 -0.5940  1.2970
#> -0.9448 -3.0318  1.7004 -1.6137 -0.4888 -1.4961 -0.7593  0.1588 -2.1847 -1.1563
#> -0.5491  0.5912  0.8245  0.0647  0.6202 -1.3016 -0.8391  1.0885 -0.2623 -1.3004
#> -1.4966  0.1992  0.4802  0.2187  2.0805  0.5669 -0.6759 -0.9965 -0.6486  2.4130
#> -0.6980 -0.2344 -1.2887 -0.4401  0.9496 -1.2936  1.5493  0.7672 -0.2972 -1.2378
#> -0.2214 -0.0452 -0.5472 -0.3342  0.7920 -0.3036 -1.1765 -1.0185  1.3437  0.6909
#> -0.8486  0.1293 -0.7977 -1.0694  1.8460 -1.3217 -1.3126 -0.5144  0.3606  0.3898
#> -1.4321 -0.8465 -0.1011  0.2428  1.3432  0.0318 -1.1483  1.1945  0.2513 -0.7157
#> -0.3922 -0.8158  0.0325 -0.0615 -1.1060 -0.3731  0.0971  0.3346 -0.3074 -0.5348
#> -0.8243  1.0059 -0.0095  0.8386 -0.0663  0.9392 -0.2222 -1.2268  0.1572  0.6377
#>  0.4369 -0.2884  0.4755  1.9107  0.4895 -0.3302  0.2694  3.2325 -0.2247  0.6381
#>  3.5936 -0.5225 -0.0640 -1.2469 -1.2571  0.5733  1.5483  0.8391 -0.5997  0.1031
#>  1.2974  1.7353  0.6055 -1.1878 -0.5442  0.9636 -0.4894  0.3725 -0.2841 -0.1067
#>  0.0644 -1.0093 -1.0834 -1.3939  0.8662 -0.8593  0.9533 -1.2881 -0.5112  0.0183
#> -0.6066  0.8545  1.0325  0.6407 -0.0062  0.1880 -2.4256 -0.5423 -2.1220 -0.5367
#>  1.4347  0.8069 -1.3746 -1.1120 -0.3535  0.3946 -0.7448  0.9888 -0.2277  0.1727
#> -0.9577 -0.2491 -0.4708  0.4141 -0.6247  1.2072  2.0651  1.3048  2.2104 -0.3496
#>  0.0680 -0.1598 -0.8394 -0.6341  0.2775 -0.4614  0.4406  2.7327  1.8559 -2.1938
#>  2.0037 -2.3621  1.1029 -1.1037  1.0978 -1.3194  2.1452 -0.9350  0.4886  1.2591
#> -1.0017 -0.1365  1.7054  0.7657  0.6491  0.4308  0.1776 -0.6424 -0.2872 -0.4108
#> -0.3118 -0.5154 -1.1514 -0.3412 -1.7262  1.2771 -0.3183  0.1961 -1.4195 -0.9683
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]