Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.3855 -0.4543 -0.8401  1.9734 -0.7900 -2.2088 -0.1929  1.3789 -0.1585 -0.7394
#> -0.2193 -0.4081 -0.4379  0.0787 -0.5914 -0.9739  0.5903 -0.8554 -0.8142 -0.1368
#>  0.1715  0.9701  1.7006  0.2585  1.0155  0.1378 -0.7396  0.5384  0.1013 -0.0368
#>  0.1479  1.7236  0.2917 -1.1338  1.5579  0.4903  0.8153  2.5014 -1.2168  0.4421
#>  1.6697  0.5886 -1.4404 -1.0921  0.8102 -0.0999  0.0218  0.0253 -0.4864 -3.2082
#> -0.3329 -1.2590  0.8420  2.3390  2.1659  0.5472  0.2457 -1.1145 -0.8289 -2.1649
#>  0.2290  1.4266 -0.8131 -1.1951 -0.9979  1.0903  0.2681 -0.1757 -0.2666 -0.5458
#> -1.9678  0.8531  0.0655 -1.0339 -0.2143  0.1589 -0.5459  1.4058 -1.1375 -1.8734
#>  1.3925 -0.1385 -0.0522  0.4310 -1.1079 -0.5211 -0.5057  1.0782 -0.3812 -2.1289
#>  1.8132 -0.2940  0.8396 -2.8751 -0.3912 -2.5232  1.6445 -1.4714 -1.2648  0.9542
#>  0.4942  0.0209  0.5598  0.1568 -0.4479 -1.2127 -1.3425 -0.1448  0.7148 -0.9169
#>  0.2573 -1.5165 -0.3159 -0.3788  0.2285 -0.6198 -0.0399  0.5520  0.2361 -0.8266
#>  0.2602  0.2911  1.9271 -1.4463  0.0140 -0.7561  1.0736 -1.0585 -1.8036  0.5527
#>  2.1307 -0.4163  1.3872  0.2693  1.1674  0.2155  1.8877 -0.2563 -0.9769 -0.8173
#> -1.1976  0.4843  1.1882 -0.0898 -1.3975  2.5240 -1.9255  1.1606  0.3969 -0.3809
#>  0.4004 -1.6984 -0.2187 -1.8174  0.7538 -0.4719 -0.3723  0.1693 -1.8006 -0.0077
#> -0.1188 -0.2447  0.9413 -1.9992 -0.7036  0.3890 -2.3434 -0.8434 -0.0386 -0.1933
#>  0.0118  0.7168  0.0989  0.4267 -0.3787 -0.8952 -0.7597 -1.9580  2.2351 -0.5996
#>  0.7285  1.3086  2.1540  0.9429  1.2535  0.5462 -0.2653 -2.4988 -0.5125 -0.7915
#> -0.3039  0.6389 -0.1132  0.5241 -1.7522 -0.0454  1.1981  1.5741 -0.5157 -0.2298
#> -0.3656  0.3313  0.0538  0.2680  0.2714 -1.1787 -0.0531 -1.2716 -0.6968 -0.5725
#>  1.2305  0.4685 -0.5470 -0.9817  1.5330  1.3028 -0.2202 -1.0228 -0.6548  1.0728
#>  0.6476  0.1885  1.0124 -1.8273  0.2485 -1.8176  0.6008 -0.3657  0.3512  0.6904
#>  0.1335  0.6813 -0.1101 -1.4604  0.2938 -1.1046 -1.0536  0.0866 -0.1793  2.1370
#>  1.4775 -1.7629 -2.1718 -0.7006  0.0513  0.3706 -0.3496 -1.6821 -0.3755  0.4156
#> -0.9006 -0.7715  0.7101  0.5130  1.0946 -0.8755  0.2501 -0.2782 -1.3604 -0.5151
#>  0.1715  1.4706 -0.4024 -0.7811  0.0831 -1.7497 -1.7736  1.1678  0.3239  0.5338
#> -0.1033 -0.3325 -0.2848 -0.6308 -0.2357 -0.9248  0.8303  0.8981  0.2075 -0.6347
#>  0.6251  0.7806 -0.8833  0.3301  1.1161 -0.4171 -0.0631  1.5196  0.3642 -0.5036
#>  1.4091  1.1566  0.4020 -0.9820  0.9955  0.2442 -1.5361 -2.2178 -1.6654  1.2589
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]