Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5334 -0.4632  0.0093  1.0846 -1.6035  1.9509 -0.7312 -0.0159  0.5071  1.2779
#>  0.3108  0.2840 -0.5015 -1.8595  0.1270  1.2633 -1.1236 -0.8556 -0.4650  0.8649
#>  1.3835 -1.6386  0.2563  2.1406 -0.5773  1.7757  0.8601  0.5117 -1.4209 -0.1620
#>  0.8490 -1.1510  1.7222  0.8962  1.2923 -0.6372 -0.7888 -0.6647 -0.5908 -0.7406
#> -1.6931  0.1554  0.6558  0.2369 -0.4930  1.1799 -0.7356 -0.9080  0.1974 -0.5452
#> -0.2418  0.5412 -1.4055  0.4672 -0.8880  0.2415 -0.6596  0.9450  0.1091 -0.5344
#> -0.1347 -0.8547 -0.5756  0.0424 -0.4182  0.1514  0.5617  2.1328  0.0292  2.1686
#> -1.1917 -1.8709  0.0113  0.0048 -1.6547  0.2585 -1.4580  0.4524 -0.8319  0.1004
#>  1.2306 -0.5251  1.0588  0.2406  0.2317 -1.3016  0.1567 -2.1757  1.4023  0.7292
#> -1.0318 -1.4067 -0.2435  0.2086 -1.6685 -1.2373  0.8094  0.6996  2.3068  1.4969
#>  1.0329  0.0258 -0.0283  0.1617  1.7774 -0.0541 -1.9794 -0.4479 -1.6432  1.5173
#>  1.1724  1.0485 -0.4446  0.3270  0.4250  1.8583  0.8510  0.3100  0.0918 -0.5958
#>  0.1185  0.6980  0.8906 -1.7989  0.5547  0.0550 -0.1931  1.6439  0.4784  1.6476
#>  1.5188  0.2187  0.6440 -0.1210  0.5835 -0.5248 -1.3718  0.4241  1.0039  0.0725
#>  1.2089 -0.4547  0.9312 -0.7146  0.5516 -0.6805  1.4587  0.4216 -0.4229 -0.5088
#>  1.4413 -0.2438 -0.9066 -1.6792 -0.7389  1.3893  0.5207  1.1136 -1.1200  0.2763
#> -0.4307 -0.3951 -1.9647 -1.9250  1.1072 -1.2915 -0.3541  0.5518 -1.3060 -1.1212
#>  0.0059  1.0868  0.0691 -0.4052 -0.4607  0.6680 -0.0419  0.0614  0.1245  0.4614
#> -0.3438  0.2761 -1.0861 -0.0018 -1.4909  0.9499 -0.7642 -0.9109  0.3572 -0.9426
#> -0.1036  0.3666  1.7784 -1.3530  1.6186  0.1902 -0.1975  0.8430 -0.5042 -1.6803
#>  1.4301  0.2043  0.2372 -0.8082 -1.1422 -0.0716 -0.4471 -0.6936 -1.4072  0.5534
#> -0.0651 -0.5265  0.6266  0.5275  0.8259  0.0789  1.2529 -1.3348 -0.0678 -0.6884
#> -0.1124 -0.4599 -1.3772 -2.6603 -1.3183  0.8416  1.1827  1.9488  0.8641 -0.2160
#> -0.8211  0.2102 -0.3975  0.3368  0.3438 -0.7168 -0.7792 -1.6385 -1.7219 -0.2699
#> -0.1901  0.2822 -0.1179  0.6320  0.1337  0.4673 -0.7095  0.3597 -0.2756 -0.3958
#> -0.5247  0.5973  0.3358 -0.2393  1.2297  2.1580  0.6487 -1.8389  0.3470  0.2440
#>  0.1454 -0.5565  1.4764 -0.7436 -2.2616 -0.0567 -1.0443 -0.6272  0.5031 -0.3253
#> -1.1525  0.2393 -1.0128 -0.7830  0.8425 -1.7457 -0.5524  0.9934  0.1562 -0.2524
#> -0.1645 -0.2407  0.5980  1.3017 -0.1980  0.6112  0.5495  0.4225  0.1752  0.6896
#>  2.4023 -2.2192  0.7398  0.3053 -1.6277  1.1779  0.1655 -0.3397 -0.6289 -0.1388
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]