Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5590 -0.2364 -0.5572  0.1645 -0.7385 -0.8933 -0.8021 -0.0786 -1.8812  1.5168
#> -2.1232 -0.0787 -0.2377  0.2164  0.4758 -1.7794  0.1381 -0.0795 -1.4380  1.4914
#>  0.3494 -0.1045 -0.8503  0.3208 -0.2343 -0.7410 -0.5402  1.4069  0.8439  0.3741
#> -1.7160 -0.6892  0.4097  0.1784 -0.9809  0.0134  1.8491  0.5097  0.3213 -0.2765
#>  1.3138 -0.8784  0.0357 -0.0716 -1.2469 -0.2904 -0.3284  0.2447 -1.5573 -1.3239
#> -0.0648  0.2933 -0.2767  0.6220 -2.2037 -0.2776  0.4336 -2.6116 -0.3070  0.9771
#>  0.2655 -0.2557 -1.3828 -0.0034 -1.1357  0.2082  0.3799 -1.9885  1.1243 -1.9956
#>  0.4618  0.5550 -0.0512  1.2824  0.4885  0.7092  0.4321  0.7518  0.2157  0.2351
#> -0.7138 -1.5237 -0.8482  0.4682 -1.0381  0.6092 -2.0178  0.5497  0.3435 -2.6630
#> -1.3205  0.8184 -0.0122  0.5106  0.5128  0.9390 -0.5763  2.1260  0.2104 -0.1683
#> -2.7907 -1.0919 -0.8878 -0.5664 -1.5916 -2.2820 -0.2678  0.6486 -0.0712 -0.2638
#> -3.0083 -0.4343  0.0459  2.6300  0.4411 -0.8674 -1.9887  0.7102  0.4308  0.7148
#> -1.9217  1.2482 -1.8160  0.7536 -1.5985 -1.3663 -0.2559  0.3924 -1.8988 -0.2164
#>  0.8836 -0.1436 -1.2788  0.8459  0.1622  0.1862 -0.1714  0.1787  0.1752  0.3416
#>  0.1308 -1.4840  0.8522 -0.2405  1.0092 -0.6081  0.8000 -0.7930  1.1227  1.9364
#> -1.6275 -2.4264  1.0145  0.6115 -0.1985 -2.5705  0.7152 -0.9471 -0.5225 -1.2016
#>  0.1531 -0.5235  0.7779 -0.6872  1.2262 -0.0033 -1.0593  1.0767  1.2254 -0.6286
#>  0.7037 -0.2973  0.2781  1.4211  0.5542  1.7566  0.8533  0.2899 -0.0535 -1.6233
#>  0.2393  0.0442  0.1037  1.6924  0.8406 -1.8219  0.6643 -1.5078  0.3830 -0.3906
#> -0.2048 -0.1731  0.5634  0.0522  1.0051 -0.4337 -0.8388 -0.2209 -1.2630 -1.4865
#> -0.8381  0.1452  0.7644 -0.0494  1.6296 -0.0766 -0.9347 -0.6471  0.1225 -0.1631
#>  0.7105 -1.6015  2.2013 -1.2427  0.7626 -1.1647  0.3083 -1.7935  0.3268 -0.5565
#> -0.6368  0.6512  0.3307  0.0370 -0.5186  1.5932 -1.2297  1.3486  0.9267 -0.0313
#> -0.4099  0.4553  0.2615 -0.1363  0.7463  0.2446  0.5611 -0.0162 -0.8797  1.2716
#> -0.6591  1.0370  0.3130  1.0563  1.4025  0.8879 -1.4733  0.4973 -0.7724  0.0098
#> -0.9349 -0.0561  0.3975  1.1734 -0.2215  0.7486 -1.7806 -0.2274 -0.6291 -0.8573
#>  0.6200 -0.6529 -0.2422 -1.0104 -0.6250  0.6688  1.5429  0.1345 -1.4696  0.6261
#>  0.5397 -0.0978  1.4959  0.0402 -0.2193 -0.0371  0.1345 -1.0026 -0.3189  1.6034
#> -0.1295  0.6180 -0.5181 -0.4519  0.0717  0.5725  0.9238  0.5060 -0.0173 -1.9520
#>  0.5988 -0.6058 -1.7493 -1.2242 -0.1056 -0.2677 -0.2822 -0.4378  0.6797 -0.8635
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]