Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6-8.8472e-01 -4.0033e-01 -1.5195e-01 -1.4994e+00 -9.6987e-01 -1.1732e-01
#>  1.9464e-01 -1.5488e+00 -2.4567e-01 -4.3664e-01  7.6762e-02  1.1616e+00
#> -7.1147e-01 -1.8987e+00  4.9688e-01 -3.4779e-01  2.2238e+00  7.8719e-01
#>  1.4677e+00  3.7710e-02  8.7671e-01  1.1649e+00  2.9920e+00  4.0102e-01
#>  1.8294e-01  4.1829e-01 -2.0973e+00 -2.9634e-01  9.1136e-01  9.0158e-01
#>  1.9140e-01  7.2275e-05 -4.9038e-02  7.4822e-01  1.5394e+00 -4.3375e-01
#>  5.0189e-03  1.1007e-01  1.2169e+00 -1.3252e+00 -2.6889e-01 -5.6510e-01
#>  2.0788e-01  5.0529e-01  9.4043e-01  8.3665e-01 -2.8430e-01 -2.1916e-01
#>  1.0229e+00 -2.5515e+00 -1.7309e+00 -1.0292e-01 -1.9851e-01  1.4316e+00
#>  6.5218e-01  6.0577e-01 -4.3111e-01 -5.0174e-02  1.9793e+00 -1.1458e+00
#>  1.9078e-01 -3.0069e-01 -1.1378e+00  1.6947e-01 -2.8825e+00 -6.9485e-01
#>  7.0402e-01  1.8397e+00 -6.5718e-01 -3.8175e-01 -2.7473e-01 -3.0698e-01
#>  2.1682e-01 -1.9032e+00  4.9944e-01  3.5088e-01 -2.0637e-01 -1.7648e+00
#> -9.6345e-03 -4.6352e-01 -9.2543e-01  1.9565e+00 -9.9749e-02 -2.2323e-01
#> -3.9649e-02 -4.0398e-01  5.8089e-01 -6.6561e-01  6.3867e-01 -6.5964e-02
#> -2.5841e-01  3.5855e+00  1.0945e+00  8.5587e-01 -5.8831e-01  4.0044e-01
#> -1.9291e+00 -4.6296e-01 -3.1665e-01  1.8301e+00  9.3512e-02  6.5643e-01
#>  4.9572e-01 -1.3910e+00  1.1083e+00 -2.2983e-01 -3.3413e-01  4.6691e-02
#> -2.6332e-01 -4.7450e-01 -1.3545e+00 -4.2371e-01 -1.3407e+00  1.4148e+00
#>  1.3818e+00  6.7067e-01  6.6911e-01  2.8020e-01 -7.0825e-01  7.8183e-01
#>  1.4150e+00 -1.5945e+00  7.2115e-01  7.7917e-01  5.6912e-01 -1.1608e+00
#> -3.9368e-01 -1.1013e+00 -6.9910e-01  5.0951e-01  2.1344e-01  1.2028e+00
#> -1.9224e-01 -1.5183e-01 -2.0041e-01  1.4610e+00 -1.1011e-01 -6.8206e-01
#> -1.3913e-01 -4.6376e-01 -4.1473e-01  2.5207e-01 -3.5026e-01  1.2345e+00
#> -8.7498e-01  2.6037e+00  1.3122e+00  1.4887e+00  1.4825e+00  1.4164e+00
#> -8.6763e-01  3.9732e-01 -1.1955e+00  6.6293e-01 -2.7335e-01  8.1446e-01
#>  9.1645e-02 -9.5765e-01 -2.8302e-01  4.9696e-01 -2.1299e+00 -1.3948e+00
#> -4.8184e-02 -6.1188e-01 -7.4467e-01  1.4127e-01  3.1700e+00 -7.6014e-02
#>  9.0651e-01 -1.8295e-01 -2.4664e-01 -1.0829e+00 -4.8900e-01 -8.0959e-01
#> -3.0459e-01 -1.9593e+00 -1.2439e-01 -2.2695e+00  3.1593e-01 -6.4624e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]