Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-2.0736 -2.2356  1.2622 -1.4643 -1.2798  1.0461  1.3260  1.0546  0.0393 -0.9550
#>  0.8623  2.3215  0.8516  0.7186 -1.1459 -0.9241  0.4748 -0.5265  1.8814 -0.5251
#>  1.4888 -1.2670  1.2133 -1.5349  0.0466  0.2905 -1.4512 -0.4290  0.8115 -0.8085
#>  0.6233  1.6818 -0.7782 -1.0963  1.4814 -0.2795 -0.4935 -0.2265 -0.8907 -0.4372
#> -1.7317 -0.8643 -0.0822  0.8667 -0.7902  0.2176  0.5778 -0.0378  0.9366 -0.9913
#> -0.6517  0.1297  0.7561 -0.8855 -2.3647  0.9547  0.7319  1.0648  0.5779 -1.6988
#>  0.5222  1.1947 -1.1481  0.2194 -0.2272  0.3148  1.1931 -0.8291  0.6673  1.4834
#>  0.0524 -1.4531 -0.6607  0.2728 -0.7966 -0.4004 -0.5875  0.0296 -0.2445  0.1253
#>  0.4169  0.6253  0.0049  0.3346  0.3653  0.8689 -2.1935 -0.5176 -0.2628 -0.7561
#>  0.6094  1.3626  0.1765  1.5554 -0.5768  0.2436  0.5929 -0.4154 -0.4913 -0.2120
#>  2.0169 -0.1146  0.1615 -0.6763 -1.6765 -1.5758 -0.2454 -1.2097  0.5146 -0.0738
#>  1.7829 -0.4266  0.9453  1.3106  0.3952  1.2857 -0.3990 -0.0428  0.6167  1.3673
#> -2.1474  0.7653  0.1779  1.7436  1.6488  1.1115  0.5322  1.6942  0.1467 -0.6468
#>  0.6806  0.2086 -0.5359 -1.4381 -0.2389  0.0672 -0.0199 -1.7995  0.1831 -1.1318
#>  0.9428 -0.0760 -1.1140  0.0130 -0.3225  0.2683  0.4044  0.2411  0.8984 -0.2032
#> -0.5337 -0.7660  0.5628 -1.6560 -1.3940 -0.0881 -0.4578 -0.7560 -1.6835 -0.1180
#>  0.5695 -1.5354 -0.8102 -0.9515 -0.6229 -0.6884 -1.2482 -0.8506 -2.3352  0.1477
#> -0.4172  0.0856  1.5203 -1.3049  0.6290  0.8912  0.8258 -0.1437 -0.3507 -0.0623
#>  0.0838 -0.2180 -1.1478  2.1659  1.6066 -0.6452 -0.3655  0.0656 -0.3317 -1.2801
#>  0.0255 -0.1770 -0.7446 -0.1273  0.4943 -0.9682  0.0210 -0.8801  0.1246 -0.1670
#>  1.0954  0.3774  0.5665 -0.6183  0.7764 -0.3501  0.0400 -0.5936  0.0898  1.6814
#> -0.6397  0.4658  1.3391 -0.8341  2.6734 -1.7750  0.2253 -0.4525  0.7237  0.5613
#> -0.5501 -0.3435  0.7053 -0.8391  1.4544 -1.0259  0.4004 -0.0441  0.7427 -0.1596
#>  1.5137  1.0437 -1.3212 -0.1631  0.6966  0.2860  1.2331 -0.0539 -2.1245  1.0175
#> -0.1339 -0.6503 -0.0619 -0.8755  0.3578 -0.6855 -0.3092  0.0894  1.7221  0.2722
#> -0.7475 -0.1345 -0.9139  0.9830  1.4703 -1.6096 -1.1908 -0.4505  0.5129 -1.0708
#> -0.5963 -0.3450  1.1448 -2.1030 -0.5700 -0.8275 -0.5956 -0.8290 -0.1465  1.3479
#> -0.3560 -0.2761  0.2232  0.1077 -1.0868  0.9911  1.2552  0.3676 -0.3263  0.7332
#>  0.1378  0.8288  0.2427  0.8751  0.8714  0.9138 -0.7823 -0.2169 -0.9701 -1.6481
#> -2.6737  1.8519  0.5953 -0.0069 -0.2886  0.7736  0.7546  0.0891 -0.0062 -0.5111
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]