Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.6096 -0.2727  0.0724  1.0577  1.5027  1.4464 -0.8674 -1.3863 -1.1056 -0.0400
#>  1.0814 -0.8696  0.7319  0.9948  0.5785  0.4148  0.4001 -0.5590  0.1508 -0.0453
#>  0.4055  1.2258  0.3633  0.8051 -0.3631  0.6155  0.0240 -1.7315  0.1497 -0.2195
#>  0.1356  0.2768  1.3805  0.0290  1.3333  0.3202 -2.1116 -0.9995  2.7115  0.0970
#>  0.5685 -1.5138 -0.5028 -1.3412 -2.9022  0.9070 -0.8049  0.8390 -0.1220  0.7069
#> -0.3575  0.6549 -0.4763  0.3518  1.0849  0.3547 -0.7915  0.7559 -1.3417 -0.2525
#> -1.4205  0.7044 -1.0800 -1.8009 -1.0332  0.7446  0.3108  0.1323 -1.2427  0.6412
#> -0.2175 -0.1733 -0.0714  2.0979  0.7151 -1.5258  2.2881  0.5692 -1.2380  0.9655
#> -0.5445 -0.1498 -1.6226 -0.3819 -1.8656 -0.9684  0.9233 -0.5635 -0.3511  0.3685
#> -0.8178 -0.1136 -0.2341  0.9790  1.1055 -0.5939 -0.9350  1.4161  0.3826  1.2866
#> -0.7126 -0.5918  2.5609 -0.3054 -1.5831  0.7884  1.7502  2.1691  0.9673  1.1642
#> -0.0150  1.7415  0.6266 -0.7486 -0.7379  0.5544 -0.4868  0.3009 -1.3712 -0.6157
#>  1.7878  0.0210  0.2124  0.0222 -1.6374 -0.3417 -1.2008 -1.0974 -0.5008  1.4841
#>  2.6177  0.6365 -1.0435  0.5182 -1.5256 -0.7900  0.6630 -0.7232  0.2491 -0.5339
#> -0.3006  0.4226  1.7916  0.6091 -0.8727  0.0377 -0.8929  1.3053  0.4657  1.5484
#>  0.2332 -0.9739  0.8179 -1.5295 -0.5942 -0.1880 -0.0968  1.2669  0.2089  0.0296
#>  2.4767 -0.5022  0.0321  1.1081  0.5474 -0.3844 -2.1666  0.0362 -0.0650 -0.8509
#>  3.7165 -0.2655  0.2463 -0.0903 -0.6691  1.5431  0.6243  0.2099  0.4488  1.9647
#>  0.1797 -0.7153  0.7028 -0.0470  1.4211  0.0143 -0.3578 -0.7562  1.4721  0.4080
#>  0.1372 -1.4950  0.6375 -0.4025 -0.7775  0.4447  1.2695  1.4779 -0.1537  0.7919
#>  0.0161 -0.3132 -1.9472 -1.3979 -1.2630  0.4382 -0.0344 -0.5976  0.7328  0.3264
#> -0.9554  0.8202  0.2654 -0.5755 -1.6804 -0.1019 -0.0687  1.3373  0.8118  0.7217
#> -1.3150  2.5731 -0.1091  0.0282  1.1437  1.4509  0.8487 -0.1658  0.0556 -1.2592
#> -0.0417 -0.5582  1.6102 -0.8126  0.6679 -1.0525  0.3107 -1.2671 -0.0970  0.5831
#> -0.1342 -1.4396  0.3814  0.8094  0.1424 -0.7143  0.2988  0.0726  2.0183  1.6787
#> -1.1651  1.4499  0.9383  0.7395  0.0954  0.5980 -0.7004  1.8863  0.9300 -2.9610
#>  0.9271 -0.0425 -1.5002  0.9810  0.2679 -1.0071 -1.0300 -0.7668  0.9343 -0.0583
#> -1.6124  0.6838  0.8979 -0.3498 -0.2924  0.5959 -0.7678  0.0844 -0.4740 -0.7376
#>  1.5498 -1.0196  1.6805 -1.1172  1.6040  1.2262  0.1922  0.6489  0.7936  0.2109
#> -1.2806 -0.9801 -1.3261  2.5984 -0.2107 -1.3936 -1.0236 -1.0711  2.2190  0.2799
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]