Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10 1.0361 -1.1980 -0.9074  0.2438  1.2970 -0.8808 -2.4754  1.5922 -1.3683 -0.4281
#> -1.2024 -0.0621  2.1911 -0.2026 -0.4072  0.0381  0.1269 -0.3578 -0.7498 -0.5935
#> -0.2195 -0.5736 -1.0645 -1.8604  0.6731 -0.0053  0.4650 -0.3320 -0.6780  0.2549
#>  0.4140  0.6105  0.2059 -0.5552 -0.3907  1.1285  0.9026  0.0411  1.2145  0.2708
#>  0.5614 -2.3975 -1.2575 -0.3108 -0.2794  1.6743  2.0728  0.7377 -0.0946  0.8359
#> -0.2710 -0.1880 -0.5095  1.2718  0.2532  0.6694  0.2744 -0.9580 -0.4326 -0.0676
#> -0.1527 -2.1829  1.3707 -0.4688 -0.0501 -0.2179  2.8604  1.0151  0.2282  0.8078
#> -1.1884  1.3570 -0.4316 -0.8795 -0.4290 -0.0224 -0.6103 -0.9245  0.8577  0.8058
#>  0.0679  1.5672 -0.0481  0.6587  0.3765 -1.6981 -0.3492  0.6421  1.3051  0.6834
#>  0.6555 -0.9592 -0.8659 -1.6092 -0.3233 -0.6799 -0.9424  0.8576  1.0260  0.2037
#>  0.9630 -0.1891 -1.0184 -0.2792  0.5510  1.9242 -0.1457 -0.7551 -0.7321 -2.8205
#> -0.9416  0.0039  1.6046 -0.1901 -0.8606 -1.7476  0.6357  0.8171 -1.4611  0.9140
#> -1.5493 -0.0609 -1.6791 -0.2779  0.6139  0.7655  1.0027 -0.3684  0.2920 -1.5101
#>  0.9997 -0.1986 -0.4838 -0.0328  1.9236 -0.1761 -0.6363  1.6047 -0.3590  0.3506
#>  0.6907 -0.8951  0.7056 -0.3721  0.4035  0.3829 -0.5299  0.0805 -0.6380 -0.5347
#> -0.9649  0.2725  0.9230 -0.2759  0.3736  1.2766 -0.6942  0.4603 -0.1781 -0.5770
#> -0.0622  0.6352  1.2644  0.0272  0.2444 -0.8734 -0.6565 -0.3844 -0.8745  0.5295
#> -0.2158  0.5127  0.1475 -1.7570  0.4212  0.4779  1.2672  1.5575  1.3698  1.3848
#> -1.8372  1.2986  1.1390  1.0905  1.9455  2.7069  0.7000 -1.6625  0.2373 -0.7284
#> -0.2560  1.1554  0.1011  0.5691  0.1927  0.3770  0.6944  1.6784  1.3589  1.1884
#>  0.3240 -0.5577 -0.6773 -1.8449 -0.3132 -1.0710 -0.7310 -0.8034  0.7414 -0.9993
#> -1.0793 -0.6805  0.7028 -0.4767 -0.4080 -1.2755  0.8421 -0.4249  0.6071 -0.6335
#> -0.4508  1.3341 -0.3183 -0.7703  0.4930  1.6592  0.8122 -0.3083  0.0123 -1.3280
#> -0.3528  0.2824  0.9835 -1.5374 -2.2997 -0.4455  0.1472 -0.4146 -0.3745  1.1510
#> -0.4445 -0.3468  0.0907 -0.3278  0.5130 -0.5102 -1.1292  0.8753  0.9249 -0.0099
#> -0.7768 -0.5092  1.0796 -0.2721  0.6003 -0.4041 -0.3285 -1.9275  0.1272  1.4284
#>  1.6899 -0.0600 -0.4581 -0.4939 -0.0477 -1.7448  2.2860 -0.9725  0.8158  0.4908
#> -0.9008 -0.6372 -0.1649  1.4479  1.5381  0.2305 -1.2941 -0.5830  0.3915 -0.4135
#> -1.2202 -0.4269  2.3943  0.6158 -0.8659  1.9931  0.7516  0.2194 -2.1310 -0.9492
#> -1.2286  0.3913  0.1137  1.0464 -1.2815 -1.7532 -1.1951 -0.6752 -0.9521  0.0146
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]