Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.9580 -1.1261  1.8106  0.3106 -3.1334 -0.3246  0.9193 -0.2221  0.5793  0.4773
#>  2.4078 -1.1398  0.1264 -0.4711  0.7422  0.6798  0.6040 -1.0212  0.0587 -0.1341
#>  1.2471  1.1580  0.4092 -0.7725 -0.7294  0.3532  0.5192 -0.4528 -1.6189  0.2195
#>  0.4440  1.7346 -1.2358  1.3150  0.8449  0.2520 -1.6396 -0.5225  0.2987 -0.6502
#>  1.1544  0.5245 -0.9665  0.9124  1.4074  0.2849 -0.5895  0.3689  0.4604  1.4776
#> -0.3253  1.7940  0.7970  0.2203  0.0357 -1.5269 -1.8035 -0.3100 -0.7809  0.9302
#> -1.2673 -0.3590 -0.2320 -0.3030  1.9733 -0.1844  0.4426  0.9905  0.4650 -1.5767
#> -1.2720 -0.0963  1.7392 -1.1354 -0.2202 -0.6040 -1.0306  2.0796 -0.8799  0.2072
#>  0.4995  1.0538 -1.3771 -1.1718  0.1269  0.3432  0.3034 -1.9233  0.1409  0.3891
#>  1.2966  0.7166  1.4008 -2.0704 -1.5690 -2.8525 -0.1688  2.1632 -0.7684 -0.6250
#> -0.5651 -0.1619 -1.8017  2.2272 -0.3155 -1.4386 -0.6944 -0.3303  0.2904 -0.6398
#>  0.7365  1.2344 -0.5776 -0.3101 -0.6875  0.6634  0.7541  0.8281  1.6580 -1.4567
#>  1.0998  0.7189  0.0721  0.8900 -0.9853  0.5812  0.3826  1.0209  1.5047  0.5774
#> -0.0120 -0.5254 -0.9757 -0.0237  0.1078  0.2753  1.0426  0.3366 -0.2162 -1.0139
#>  0.4403 -1.0830 -1.1002 -1.5484  1.5295  1.5832  1.9531 -1.0899  0.1983  0.1385
#> -0.7533  0.3803 -0.8699  2.5585 -0.8506  0.4786  1.2004 -1.4586  0.2647 -1.8708
#> -0.8915 -0.2334 -1.0511 -0.2010  1.8928  2.4436 -0.9191  0.7839  1.0820 -1.0015
#> -0.4647  0.0166  0.4149  0.3646  0.3608 -0.1841 -1.6021 -1.2637  0.2452  2.1547
#>  1.3739  0.8001 -0.0060  2.0343  0.8105  0.0556 -1.5445  0.5977 -0.2782 -1.7026
#>  0.7801 -0.1559 -1.3286 -0.2723  1.9274 -2.7327 -0.6106 -0.1725  1.4257  1.1940
#>  1.1836  2.8550  0.2715 -1.5384 -0.3679  2.4704 -1.2590 -0.2284  0.1325 -0.1392
#>  0.7319  0.5438  1.4756 -0.4909  0.1269  3.0689  1.2187 -1.0106 -1.2331  0.0397
#> -0.2423  0.0874  0.4846  0.2596  0.5069 -0.3744  0.3092  0.7333  0.2814  0.6157
#> -0.1984 -0.8022 -0.8024 -0.4150  0.0606 -0.8769  0.0831  2.1594 -0.6612  0.3973
#> -0.5582 -0.1169 -0.1361 -0.8960 -0.9141 -0.0983  1.1869  0.0745  0.0590 -0.4819
#>  0.4768 -0.4493 -0.5310 -0.0202  0.5890 -1.9234  0.8073 -0.2250  0.3825 -0.3503
#> -1.3807 -0.3672  2.2090  0.3087 -0.6447 -0.2703  2.0529 -3.1078  0.9577  1.0523
#>  0.8376  1.7779  0.0640  1.5179 -0.3918 -0.6445 -0.6303 -0.7990 -0.3156  1.9775
#> -0.3799 -1.9919  0.4431 -0.4146  0.0475 -0.6406 -1.6695  0.3046  0.1808 -0.6004
#> -1.7942 -0.4462  1.2452 -0.5014  0.5018 -0.0652  0.9774  0.5422  0.4588 -1.4721
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]