Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.9277 -0.1792  0.2179 -0.2463  0.2416  0.5353  0.1802  0.6120 -1.3491  0.1681
#>  0.6295  0.0531  0.4936  0.8508 -0.1407 -0.3013 -1.9580 -0.0740  0.6521 -1.9641
#> -0.3133  0.5093 -0.0469 -0.2271  1.6740  0.8873 -1.3648  0.3614 -0.2601  0.2113
#> -0.3257  0.6391 -0.7412 -1.9798  0.2254 -0.1610 -0.0448 -1.6045 -1.1090  0.7758
#>  1.3719  0.0933  0.1142  0.2071  0.7085  0.6618 -0.2836  2.0744  0.7587  1.9771
#>  1.6124  0.2718  1.0773  1.7089  0.5655 -1.7005  0.3064  1.1482 -0.6508 -0.3177
#>  0.0069  0.1001 -0.3821  0.2470  0.1717 -0.7779 -1.0777 -2.1659  1.5389  0.4716
#> -0.9933  0.8853 -0.6666 -0.0437  0.0645 -0.0034 -0.6426 -0.9600 -0.7979 -0.7670
#>  1.6464 -1.1996 -0.4439  0.2463 -0.9263 -1.4183  0.3487  0.1123  0.8961 -0.3522
#>  0.5171  0.2836 -0.6524 -1.2268  0.9747  0.5212  0.9161 -0.2090  2.2757  0.8560
#>  2.7969  2.3391 -0.3137 -0.6472 -1.2445  0.1603 -0.5844 -0.2403 -1.4849 -0.9108
#> -0.2334  0.7239  0.9104 -0.3236  1.4191  1.4305  0.3794 -2.9808 -0.3140 -0.3216
#> -0.6686  0.1182 -1.5521  1.3043 -0.4829  0.5892  0.2251  0.6197 -1.2946 -1.1205
#>  1.7142  1.0154 -0.3657  1.7717 -0.0822  0.8567  0.9006 -0.3479 -0.6398 -1.8893
#> -0.7114 -0.9681  0.9482 -0.3142  1.9025  0.5353  0.7397  1.2482 -0.0501 -0.5438
#> -1.0908  1.1490  0.2471  1.8863 -0.5432  0.0059 -1.2027  0.5196 -0.8125 -0.5405
#>  0.6967  0.2506  0.4276 -0.4398  1.0508 -0.7302 -0.8575  1.7893 -0.8130 -0.8973
#>  0.4775 -0.1023 -0.9870  0.9594  0.3883 -1.0429  1.1497  0.9334 -1.3531 -1.2091
#>  0.2566 -1.5733 -0.7967  0.0363 -0.4237  0.6794 -1.8367  2.3343  1.3831  0.5670
#>  0.0677  1.4182 -1.9981  1.8205  0.1410  1.6484  1.9385  0.4882  1.3309 -0.7762
#> -1.2475 -0.4518  0.0621 -0.1195  0.5055 -0.7215 -1.0379  0.3864 -0.3765 -0.6058
#>  0.9972  1.1674  0.0615  2.5370  0.3808  0.2562 -0.8149 -1.0963  1.2163 -2.0467
#> -0.2732  0.1047  0.0741  0.0288  0.3636  0.6592 -0.2734 -1.2005  1.1746  1.3074
#>  0.5619  1.5591 -0.4174  0.2134  2.0176  2.1647 -0.9579  0.2399 -0.7442  0.2466
#> -0.7854 -1.9879 -0.3623 -0.3940  0.0503 -1.2565 -0.6354 -0.1707  2.3539 -2.0027
#>  1.0925  1.4953 -0.6264  0.0376  1.2143  0.2582  0.5820  0.5838 -0.3933  0.4519
#> -0.4168  1.5903  2.0495 -0.2352 -0.0611 -0.6295 -0.6564  0.2139 -0.3466 -0.5436
#>  1.0398 -0.0590 -0.4196  1.1572 -0.0551 -0.2568 -1.6172  0.0094 -0.5509 -1.4499
#>  0.7486  0.6533  1.1112 -0.7846 -1.6528  0.5263 -0.7257  1.4550  0.8849 -0.0725
#> -0.4649  0.3245  0.9102  1.1311 -2.1225  1.3142 -0.9309 -0.5222  0.8281  1.6385
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]