Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.1446 -1.2046  0.3175 -1.6231 -1.3093 -1.9641 -0.7134 -1.3841  1.0129 -0.2874
#> -0.8513  0.9143  0.2415 -0.3014 -1.5596  0.4239 -0.8891  0.3002  0.4267 -0.7168
#> -1.7595 -0.3786  0.8240 -2.5340  1.3537  0.6653 -0.7525  0.6341 -0.0915  1.3421
#>  0.9787  2.1558  1.2778  0.5547  2.1412 -0.0902  1.2097 -0.9905 -0.7213  0.6415
#>  0.3481 -1.5386  0.6508  2.4082  2.0326 -1.1309 -0.4121  2.2889 -0.5743  0.1940
#> -0.1455  0.7866  0.6662 -0.6722 -0.5025 -0.2567 -0.2762  0.1793  1.1883  0.0200
#>  0.8949 -0.7608  1.3680  0.1382  2.0480 -0.6691  0.4275  1.4965  1.5030 -1.3147
#>  1.4432 -0.9049  0.1954  0.3881  0.0580  0.5010  1.4760  0.7605 -0.9195  0.6172
#> -0.3831 -2.5840  1.4668  0.4905 -1.0726 -1.5919  1.6594  0.5889 -0.7125 -0.0496
#> -1.8354  0.9521 -0.6733  0.1595 -1.4145 -0.1860 -0.6129 -0.4214 -0.6726 -1.3302
#>  0.6533 -1.2068  0.4497 -0.6178 -0.2439 -0.2325 -1.2528  0.3201 -0.0661  0.3447
#>  1.4797  0.3638 -0.1592 -0.3400 -0.6899 -0.7991  0.1189  0.3799 -0.4680 -0.2748
#> -1.4494  0.9318 -0.2157  0.4252  0.9538 -1.8129 -0.3713 -0.1984 -0.5385  2.2808
#>  0.3317 -0.1057 -1.1168  1.5906 -0.6077  0.5284  0.8790 -0.3388  0.6557  0.0005
#>  1.1487 -0.2128  0.5956  0.4275 -0.1850  1.2959 -0.0334 -1.3474  0.4265  1.1478
#> -0.6272  0.0042 -1.0286  0.9406 -2.1592 -1.3763  0.3055 -0.6253  0.0611 -0.6112
#> -0.0410  1.1670 -1.1947 -1.0080  0.2556  1.8149 -0.2229 -0.8631 -1.1082 -0.5032
#>  0.3051 -0.9929  0.8472 -0.2365 -0.9901 -0.2384 -1.1346  1.0507  0.6050 -0.5856
#> -0.1450 -2.3061  1.3819  1.4648  1.9290 -0.5731 -0.2324  1.3009  0.3303 -0.4631
#>  0.2940  0.0355 -1.2425  0.2330  0.6550  0.2193  1.2719 -0.3231 -0.7537  0.4651
#>  0.3185 -0.7129  1.2485  0.7906 -2.0063 -0.3321  0.5172 -1.2955 -0.6456  1.0639
#> -0.5053  1.2001 -1.1616  0.7045  0.9211  0.7599  1.2579  0.8458 -1.7484  0.3451
#>  0.0163 -1.0023 -0.2546  0.9610  1.7033  0.4802  0.1059  2.3040 -2.0380 -0.4114
#>  0.6074  0.9189 -1.0544  0.2199  0.4760 -0.5029  0.2783  0.1734 -0.6980  0.2167
#>  0.1194 -0.5014  1.7120  0.1019 -1.0228  0.3533 -0.0281  0.0499 -1.4080  0.9311
#> -0.5573  0.1562  0.2097  0.2225  0.9575  0.4704  2.0262  1.6831 -0.1633 -0.7248
#> -0.2636 -2.2269 -1.3689 -0.5623  0.5695 -1.1829  0.0030  0.2848 -0.3044 -1.6063
#>  0.0543  1.4966 -0.6224  0.2864 -0.6113 -0.5486 -0.0412 -0.8196  1.6035 -1.1047
#>  0.4767 -0.5955  0.2740  0.1918 -0.0195 -0.9392  0.1737  1.7597  0.4702  1.4255
#>  0.2253  0.8316  0.1466  2.5187  0.0147  0.7056  0.0544  0.7573  0.3520 -1.1030
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]