Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.1699  1.5533 -0.4137 -2.4596  0.7189  1.0131  0.0941 -1.6685 -0.2459 -1.1407
#>  0.8241  1.8014 -0.9524  0.3898 -0.1514  0.1756  0.0199  0.1460  3.5063 -0.2040
#>  0.8546 -0.0034 -0.5074  1.0672 -0.2494 -0.4699  2.3711  1.1363 -2.1690 -0.8843
#> -1.3474  0.0220  0.6070 -2.0462  0.1496  0.9735  1.9897 -1.9635 -0.8450  0.0816
#>  1.9447 -0.3131 -0.7914  0.9066  2.0279  0.1751  1.0487 -0.3628  1.2472 -0.4685
#> -0.4345 -0.3236  0.2240 -0.5447  1.7935 -1.4066 -0.8373 -0.7211  0.4692 -2.2241
#> -2.0337  0.4207 -1.2365 -1.3248  1.1468 -0.8984 -1.4518 -1.2825 -0.3760  0.0629
#>  0.8811 -0.4235  0.1386 -0.3038 -1.3105 -0.4973 -0.4517 -0.2440 -0.3175 -0.1411
#>  0.7093 -0.3149  1.3885 -0.9594  0.7190  1.1657  0.1119 -0.7422  0.9405  1.5739
#> -0.3352 -1.5445  1.8726  0.8863 -2.4335 -0.9475 -0.6436  1.0696  1.5535  0.5361
#> -0.2304  0.6305 -0.0594  0.3466 -0.5098 -0.8826 -0.2511  0.3245  0.0627 -0.4049
#>  1.3604  0.3638 -0.2189 -0.4788 -0.3187 -0.8216  0.4578 -0.2138 -0.7871  1.6054
#> -0.7008  0.6450 -0.6686  1.7657  1.6437 -0.3550 -0.5761 -0.3195 -0.7046  0.1273
#>  1.0317 -0.1924  0.6671 -2.0312  0.6186  0.8654  0.6906 -0.7041  0.6468  2.9131
#>  0.1849 -0.1611 -1.7880 -1.2183  0.7362  0.6190  1.2077 -0.3207 -0.3400 -0.3838
#> -0.3827  0.0328 -0.3225 -0.3312  0.8586 -0.0479 -0.0049  0.5429 -0.8740 -0.6156
#>  0.3751  1.5333  1.9477  0.5823 -0.1211 -0.1027 -0.4531 -0.3452 -0.3004  0.5509
#>  0.0080  1.5177 -0.9201  1.1675 -1.8494 -0.5786 -1.0089 -0.6927 -0.0623 -1.7994
#>  0.6138  0.5439 -0.1671 -0.7403  0.1421 -1.9122 -0.5079 -1.3127 -0.3379  0.0768
#>  1.0020  0.1855 -0.7677 -0.1771 -0.5303  0.4235  0.5982  0.7657 -0.1055  0.5725
#> -0.9990  0.1511  0.6320 -1.0180  0.5719 -0.1269 -1.1602 -0.3976 -0.2742  0.0539
#> -0.8281  0.2376 -0.2363 -0.6848 -0.7419 -0.0790  1.3860  1.4123  0.8947 -0.4203
#> -1.6042 -0.0184  0.0767  0.5705 -0.8631  0.5832  1.0559 -0.4728  0.6252  0.0461
#> -0.6890  1.3596  0.1460  0.0429  0.2610 -1.0927 -0.8285 -1.7341  0.6428  0.3341
#> -0.0480 -0.8863 -0.4551 -0.6433 -2.2735  1.3812 -0.6733  0.8919 -0.7239 -0.9881
#> -0.8117  0.7823  0.7713  1.2778 -0.7081  2.2025 -0.1982 -0.3421 -0.4313 -0.0339
#> -0.9319  0.7409  0.4698 -0.1857  1.2203  1.8840 -0.8386  0.7439 -1.9750  0.2705
#>  0.1694  0.5918  0.4138 -0.0656  0.2031  0.8454 -0.6811 -2.1066 -0.9501  0.1660
#> -0.9922 -0.6279 -0.1254 -0.0160  0.7249  0.4363  0.0036 -0.0866  0.2100  0.8117
#>  0.7209 -0.6232 -1.8406 -1.2126 -0.5929  0.2072 -0.2066 -0.7693  1.5433 -0.6192
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]