Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.5045 -0.4002  1.7759 -0.7853  0.1668 -1.4098  0.9211 -0.2241  0.8476  1.8674
#> -0.7598 -0.0506  0.4474  1.1997 -0.0353 -0.3416 -1.1897  0.2150  0.2741 -0.1265
#> -2.2563  0.6435  1.9150 -2.6043  0.6064  0.3344  0.3131 -0.0645 -1.4562  0.9472
#>  0.3185  0.8198  0.9091  0.1601 -0.4200  2.4613  0.5367 -0.2993  0.2304 -1.1627
#>  0.4714  1.4487 -0.5784 -1.2552 -0.5048 -0.9665  1.4587 -0.3152 -0.0219  0.0979
#> -0.9275  0.2850 -1.1754 -0.1172 -0.0330 -0.0521 -1.3246 -2.9119  1.3570 -0.0952
#>  1.8422 -1.6535 -1.4232  0.4260  0.3816 -1.7270 -0.6742  0.5413 -1.8367  1.3267
#> -0.5210  1.2643 -1.2415 -0.6602 -0.9010  0.5461  0.0364  0.8201  0.3271 -0.5556
#>  0.2441  1.0758  1.1872  0.3221  0.9420  1.8275  0.8222  0.9991  0.0395 -0.1150
#> -0.9500 -0.4591  0.3221  1.0250  0.5865 -1.6399 -0.6913  0.2891  1.3992 -1.1814
#>  0.9136  0.9414  2.5857  1.0156 -0.5886  0.3046 -1.5915 -0.8250 -0.9144 -0.2522
#> -0.6764  1.7675  0.3732  0.0113  0.3696  0.1593  0.2345 -1.2505 -0.5961  0.7272
#>  1.2347 -0.8945  0.5387  0.7992 -0.3559  0.5745  0.5165  1.8083 -0.5144  0.7888
#>  1.3060  0.7278  1.4561  1.3491 -0.9145 -1.7944  1.2885 -0.7632  0.1524  0.0412
#> -0.0725  1.5239  1.6622  1.5799  0.9827 -0.9430  0.6628  0.0507 -0.7450 -0.8654
#> -0.3283 -1.7380 -1.0273 -0.9723  0.8844 -0.5743  0.4331 -0.5446 -0.8103  0.6611
#>  0.9404 -0.5049 -1.0255  0.1334 -0.0729 -1.6659 -0.0484 -0.1074 -1.4231 -0.1932
#> -0.1523 -1.3774 -0.1686 -0.0504 -1.3666  0.0797  0.7661 -0.4100 -0.6007  1.0779
#>  0.3765  0.6330 -1.2454  0.3227  0.1528 -0.4632 -0.3900  0.3611 -1.3361  0.5767
#> -1.2159 -3.2059 -0.5772 -1.2044  0.0815  0.6847  0.1292  1.0742  0.4842 -1.1849
#> -0.7448 -1.5809 -1.2317 -0.4949  0.8216 -0.3865 -0.6967  2.2884  0.0865  1.7775
#>  0.1076  0.6027  1.2074  0.3712 -0.2180  0.8103 -0.2391 -0.4738  0.8745  0.4075
#>  0.0097  0.1396 -1.6664  1.6886 -0.1110 -0.9694 -1.1662  0.3233  0.1972 -0.3615
#>  0.9164  0.7100 -0.2281  0.3035  1.2739 -0.1518  0.9879 -0.9496  0.4664  0.4681
#>  2.2574 -1.5745 -1.1109  0.0783 -0.7624  0.3434 -0.7830 -0.3791 -0.3101 -0.9267
#>  1.2397 -1.1554  0.1615 -1.0606 -0.8609  1.5436 -0.7361  1.2804  1.2131  0.6386
#> -0.6234 -0.7386  0.6471  1.1570  0.1912  1.3614 -0.9051  0.5309 -1.3100  2.0266
#> -0.3220 -1.1383 -1.1512  2.9434  1.2777 -0.7452  0.1049 -0.2645  1.0053 -0.6117
#> -0.0860 -2.3207 -0.8466  0.7563  0.4291  0.9821 -0.3399  0.0699 -0.5578  1.1938
#> -0.0678  0.1237  0.5555 -2.3742  0.9795  0.1865  0.6203 -0.0339  0.2140 -0.9419
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]