Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.4330 -1.8921 -0.1441 -1.3824 -1.3457 -0.7202  0.6900  1.2996 -0.7539  0.2955
#> -0.4918 -0.1268  0.0118  0.4142  2.5847 -0.1882 -0.1920  0.3169 -0.6201  1.0997
#>  1.1037 -0.2637  1.4306 -0.6772  0.4739  0.4784  1.8208  0.9705 -0.0234 -0.6882
#>  1.1700  1.5329  2.8513 -0.1817 -0.4780  0.5494 -0.9981 -0.9330  1.1203 -1.3259
#>  1.1289 -0.5493  0.8321  1.3614  0.1689  0.3958 -0.2694 -0.0392 -0.7387 -0.5233
#>  0.1423  2.1963  0.6244 -0.5211  1.3169  0.5532  0.4291  0.0792 -0.3031 -0.7163
#>  0.4403 -1.7606  0.1165  2.0616 -0.2989  0.2593  0.2222 -0.7113 -1.2455 -0.0023
#>  0.2327  0.8366 -0.0539  0.0398 -0.5351  0.9806 -1.8297 -0.6520 -1.6177  0.8427
#>  0.2767 -0.2341  1.8845  0.6139  0.3445 -1.8921  0.8093  0.6996  0.5466 -1.3906
#> -0.2347  0.4704 -1.4342  0.9955  1.1113 -1.7040 -1.0241 -1.4551 -1.5525  0.7388
#> -0.7497 -0.2593 -1.2829 -0.6543 -1.0133 -0.2937 -0.9980 -0.3892  1.2521 -0.8429
#>  1.8277 -0.7253 -1.2342  0.4355 -0.0348  0.3199  0.4052  0.7602  2.2813  1.5749
#>  0.4095 -1.6084 -0.6719 -0.2638  0.0609 -0.5611 -1.3313 -0.2960 -0.3641  1.6900
#>  1.1324 -1.0686  1.2127 -1.5507  2.0606  2.3380  1.5868  0.6698  1.0403  2.5035
#> -0.8654  0.6030  1.7872  1.3349  0.5791 -0.8087  1.0364 -0.0357 -0.1060 -0.3406
#> -2.5149  1.8876  0.6846 -0.6867  0.5203  0.2588 -2.1876  0.5757  0.3527 -0.0731
#>  1.0428 -1.7219  0.7053 -0.3674 -0.2491 -0.4107 -0.3945  0.5257  0.3547 -0.7139
#> -1.0953  0.2790  0.2251 -2.2802  0.9643 -0.1325 -1.6714 -2.4793 -0.1913 -0.3368
#>  0.1232  3.0193  0.3132  1.3036 -0.7892  1.4422 -1.0116 -0.9539  0.3021 -1.0655
#>  1.0767  0.3717 -0.3937  2.4061 -0.3996 -1.8497 -0.3357 -0.8696 -0.1669  0.1921
#>  0.1887 -0.5299 -0.3269  0.2663  0.9708  0.9727  1.8692 -1.6835  1.0086  0.2893
#> -1.3690  0.1780 -1.0049  0.3403  1.7507 -0.5486 -0.0928 -0.6972  0.2756 -1.2077
#> -0.1366 -1.0859 -0.1714 -0.9919 -0.4837 -0.6280 -0.3676  0.1301  0.3840  0.9927
#>  1.0433  1.1778 -2.2592  0.5711  0.7923 -0.5793  1.1527 -2.0405 -0.5136 -0.5696
#> -1.4179  0.0475  0.2496 -1.7225 -1.2956  1.6498  0.9363 -2.2297  0.0198 -1.1325
#>  1.3327 -0.0196  1.6129  2.6128 -0.5770 -1.1447 -0.5565  1.5048 -0.7248 -0.5380
#> -0.1372 -1.1326 -1.2099  1.3321 -0.5621 -0.1388 -1.1137  0.9657 -0.6639  1.5588
#> -0.7754 -0.5595 -1.9937  0.1936  1.9125 -0.7936  2.0338  0.5540 -0.5088 -0.3299
#> -0.7994  1.4616  0.2722  0.0178  1.0688  0.5812  0.5547  0.8269 -0.3761  1.9235
#> -0.4806  1.6859  0.2360 -0.1650 -1.1628 -0.3671 -1.0821  1.0746  0.5914 -1.0958
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]