Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 1.8248e+00 -1.2706e+00 -1.5853e+00 -4.4139e-02  8.4891e-01  5.9313e-01
#> -2.4153e-01  1.1285e+00 -5.2100e-01 -1.0130e+00  1.2574e+00  2.8503e-01
#>  4.1917e-01 -1.5406e-01  5.3426e-01  5.8464e-01  1.3316e-01 -1.1253e+00
#>  1.3376e+00 -5.3641e-01 -6.9663e-01 -1.0225e+00  1.0634e+00 -3.4667e-01
#> -5.5487e-01  5.6860e-01 -9.7142e-03 -5.2407e-01 -9.5745e-01  2.3198e+00
#> -1.1235e+00  1.6753e-01 -1.5625e+00  1.3235e+00  1.3538e+00  3.0196e-01
#>  4.6384e-01 -8.1885e-01  1.6457e+00 -2.1553e-01  1.0624e+00 -6.5389e-01
#> -1.0159e+00  6.5283e-01  2.4861e+00 -4.4028e-01 -1.2049e+00 -8.3160e-01
#>  5.7196e-01 -7.4011e-01  3.7179e-01  3.3921e-01  1.1632e+00 -7.0238e-01
#>  4.4346e-01  1.6632e+00  4.0052e-01 -1.1279e-01 -4.9038e-01  1.0492e+00
#>  3.0918e-02 -9.5427e-01  3.8748e-01  1.2594e+00 -1.6396e+00 -8.9338e-01
#>  3.9442e-01  1.1386e+00 -3.8945e-01 -3.7740e-01 -2.8718e-01  1.0348e+00
#> -1.2508e+00 -3.0983e-01  9.2323e-01 -2.1572e-01  9.4905e-01 -1.7774e-01
#>  2.3347e+00 -2.0130e-01  1.1691e+00 -1.0441e+00 -4.1854e-01 -9.3692e-01
#>  1.6288e+00  1.0786e+00  1.6056e+00 -2.0909e+00  2.0420e-01  6.3690e-01
#> -1.9901e-01 -1.2826e+00  8.2229e-01 -1.4813e+00  9.5418e-01 -8.7323e-01
#> -1.7978e+00 -1.6257e+00  1.2244e+00  1.0038e-01  1.1381e+00  1.6045e-01
#>  1.1047e+00  5.1074e-01  3.9457e-01 -2.0251e+00  1.1194e+00  6.3876e-01
#>  6.2108e-02 -2.0526e-01 -5.5655e-01  1.2825e+00 -6.5619e-01  1.5635e+00
#>  3.1438e+00 -2.0010e+00  1.3466e+00 -6.7440e-01  3.8966e-01 -6.4407e-01
#>  5.7157e-01 -4.8686e-01 -9.7372e-01  3.0966e-01 -7.2873e-01  1.6520e+00
#> -2.1327e+00  1.6413e+00 -2.1878e-01 -1.8776e+00 -2.4760e-01 -2.9978e-01
#> -1.6280e+00  1.7106e+00  2.9770e-01  1.3303e+00 -7.4196e-01  1.4170e-01
#>  9.7490e-02 -6.0474e-01  8.7572e-01  5.2142e-01 -8.9142e-01  1.0763e+00
#>  9.3734e-01 -8.1704e-01  2.8056e+00 -7.4521e-01 -2.5064e+00 -6.7113e-01
#> -9.3881e-01  5.6218e-01 -8.0278e-03 -1.2915e-01 -5.4601e-01 -5.5259e-01
#>  1.4783e+00  3.3465e-01  8.4035e-01 -3.4535e-01  4.3649e-01 -1.6454e-01
#>  7.9215e-01  1.8306e-01  9.2591e-01  1.2080e-01  2.6875e-01 -1.4580e-01
#>  1.7297e+00 -2.0610e-01 -7.1148e-01 -6.4294e-01 -1.2101e-01  6.6118e-01
#>  1.2201e+00  4.3976e-01 -1.5298e-01  9.2202e-02 -2.1288e-01  5.7299e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]