Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 9.2166e-01 -7.9217e-01  2.6899e-01 -4.5523e-01 -4.2342e-01 -2.3663e+00
#> -2.6468e-01 -1.2120e+00 -1.5176e+00  1.3018e+00 -1.5982e+00 -2.7961e-01
#> -1.7090e+00  6.1562e-01 -1.2680e+00 -4.7253e-01  5.7324e-03 -3.2414e-01
#> -8.9474e-01  6.5477e-02 -6.7171e-01  1.6349e+00  7.9636e-01 -4.2986e-01
#>  6.6313e-01  1.1998e+00  4.9758e-01 -1.2057e+00  3.1903e-01 -1.9457e+00
#>  7.5140e-01 -7.1105e-03  6.7606e-01 -3.9369e-01  2.1146e+00  2.3982e-01
#>  3.7138e-01  5.4038e-01  6.2436e-01  5.2566e-01  2.7963e-01  5.2208e-01
#>  2.5634e-03 -5.0905e-01 -3.3832e-02 -7.0743e-01  7.4310e-01  2.1905e-01
#>  4.2406e-01 -5.8126e-02  2.2375e-01 -4.9757e-01  6.8202e-01  7.0496e-01
#>  1.7241e+00  8.6947e-01 -3.1833e-01  4.0105e-01 -2.0129e+00 -1.0033e+00
#>  1.0134e+00 -1.6399e-02 -4.2443e-01  6.3757e-02  7.8602e-01  8.0261e-01
#>  1.6963e+00  1.3552e-01  2.8457e-03  1.4368e+00 -4.3993e-01  1.9762e+00
#>  1.9972e+00  9.7412e-01  5.9964e-01  6.7886e-01 -3.5946e-02 -9.7553e-01
#>  8.9785e-01  5.2435e-01 -5.0351e-02 -1.5689e-01 -1.0840e+00  2.2349e-01
#> -9.2000e-01  3.2239e-01 -1.8287e-01 -3.0299e-01  2.3440e+00  1.5967e-01
#>  3.1069e-01 -3.5003e-02 -5.4037e-01  6.6791e-01  9.7549e-01 -7.0285e-01
#>  1.3502e+00 -2.3254e+00  1.9680e+00 -2.2459e-01  4.4291e-01 -3.5885e-01
#> -1.3499e+00 -6.7738e-01  3.6624e-01  1.5651e+00 -6.5311e-02  7.8574e-01
#>  7.6447e-01  1.1603e+00  1.7779e-01  5.9002e-01 -8.0502e-02  4.8680e-01
#>  8.0811e-02  1.1616e+00 -5.8251e-01  2.3126e+00 -2.1645e+00  3.5136e-01
#>  8.4026e-01 -3.8972e-01 -1.2662e+00  1.8030e+00 -3.9894e-01  1.2608e+00
#> -2.9142e-01 -8.1199e-01 -1.1618e+00 -8.1297e-01 -2.2860e+00 -3.1935e+00
#> -5.3373e-01 -7.0412e-01  3.0181e-01  8.0873e-01  8.4280e-01  1.3024e-01
#>  4.4640e-01 -4.9665e-01 -8.3985e-01  1.3344e-01  1.2401e+00 -5.7982e-01
#>  9.6006e-01  4.0588e-01  1.0525e+00  3.6256e-01  4.5970e-01  2.1800e-01
#> -2.5629e+00 -7.4858e-01  1.1203e+00  2.1951e+00  1.0474e+00  1.2663e+00
#>  1.3158e+00  6.9944e-01  5.5850e-02 -5.6011e-01  1.1851e-01 -1.8072e-01
#>  1.0490e+00 -7.8496e-01 -1.6370e+00 -5.7927e-01  1.3318e+00  2.2185e-01
#>  8.5922e-01 -2.4415e+00 -7.0393e-01  6.6104e-01  1.4345e+00 -8.2711e-01
#>  5.9395e-02 -5.3903e-02  1.3617e+00 -1.4729e+00 -2.1015e+00 -1.2416e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]