Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.4578 -0.4548 -1.2083 -0.5970  0.3965  2.0120  0.0990  0.2284 -2.7728 -0.7722
#> -0.1949 -1.4631 -1.8953 -0.2237 -0.1655  0.3125  0.5836  1.4138 -0.1257  0.3176
#>  0.7349 -0.6274 -0.6315  0.4071 -1.5777  0.0912 -0.3542  1.2392  1.0599 -1.5604
#> -1.3438 -3.2010  1.3221 -0.5466  0.4204 -1.7451 -0.2982  0.1389 -1.4212 -1.9416
#>  0.7364  0.6393  1.6319 -0.4021  1.8645 -0.7883 -0.1787 -0.5846 -0.2745 -0.8621
#>  0.7529 -0.1903  1.6405 -0.9355 -0.8126 -0.8070  0.4026  0.4319 -0.7822 -0.0068
#> -0.1367  2.1302  0.5883 -0.3870 -0.3763 -0.4534 -0.7411  0.1954  0.2641 -0.1170
#> -0.7885  1.2522  0.8538  0.9704  0.5034  0.8361  0.5513 -1.8710 -0.1449 -0.4389
#> -0.2578 -2.0845  0.4732  0.0576  0.7534 -0.3655  0.8198  1.2365 -0.3359  0.0708
#>  0.4558 -1.6557 -0.3209 -0.8544 -0.1277 -1.8947 -0.7437  1.1934 -0.4000  0.8107
#>  1.5440  0.1354  0.8189 -0.4843 -3.1480 -0.3623 -1.0612 -0.3113  0.6473  0.0606
#> -0.4112  0.1974  0.7473  0.1867 -0.9689  0.2838  0.2869 -0.5587 -1.0180  1.3343
#> -0.9425 -2.7033  0.3205  1.5293  0.0016  0.1024 -0.5986  0.7902  0.4446  1.6191
#> -0.8741 -0.6261 -0.0973  0.0821  0.3816 -0.0748  0.0109  0.3708 -1.2154  0.1154
#> -0.4070  0.7324  0.3434  1.0481  1.8080  0.5134  0.1126  0.5052  0.8548 -0.3828
#> -0.4103 -1.8559  0.2691 -0.1813  1.3297 -1.3901  1.1682  0.5234  0.2264 -0.0103
#>  0.3954 -0.3229  1.0465 -0.3612  0.4764 -0.4867  1.0952  0.8115 -1.5440  0.9424
#> -0.3630 -0.9486 -1.4452  1.4200 -1.0655  0.0363  0.7198 -0.5126  0.0681  0.4398
#>  0.5893 -0.6091 -0.7577 -0.1870 -0.1404 -0.3032 -1.8522 -0.5393  0.8637 -0.4762
#>  0.6686 -0.5519 -0.0725  0.0595 -0.5537  0.8001 -0.3707  0.8942 -0.4036 -1.0883
#> -1.8121 -1.1323 -0.0809  0.8241 -0.6153  0.6884 -0.7666  0.2723 -0.6649 -0.2660
#> -0.5051  0.1616 -2.1290 -1.1794 -0.2174  0.6034 -0.0683  0.1020  0.4617  0.6578
#> -0.9796  1.0403  0.5927 -0.6581  0.3871  0.6433  0.3656 -0.0757 -0.4496  1.1265
#>  0.6587 -1.5341 -0.1525  0.1083  0.5461 -1.3729 -0.8416 -0.1438 -0.3383  0.4705
#> -1.8959  0.2393  1.2710  0.5093  1.5028 -0.0588 -1.8892 -1.6363 -1.3785  1.5198
#>  0.9604  0.6144 -0.6566  0.9232  0.5705 -0.0045  0.3050 -2.2354 -0.2019 -1.1938
#> -0.3858  0.5868  0.4752  1.0257 -2.8645 -1.3808  1.2952 -1.7634  0.7721 -0.2345
#> -0.5663 -0.7885  1.1858  1.0874 -0.5004  0.2091  0.0717 -0.2574  0.1462  0.0906
#> -2.4526  0.6041  0.5195 -0.0278 -1.1116  1.2559 -1.1417 -0.6204 -0.8046 -1.1136
#> -1.7843  0.9238  0.1979  0.0594 -0.0903 -0.3443  0.0087  0.6176 -0.9640 -1.2027
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]