Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.1235  1.7019 -0.5808  1.8700 -3.0295 -1.7590 -1.1853  0.9531  1.4527 -0.1107
#> -0.6722 -1.2512  0.9779  0.7474  1.1876 -1.0439 -1.3444  1.7506 -0.2400 -0.6835
#>  0.5437  0.6921  0.8640 -1.0021  0.9375 -1.7828 -0.3029 -0.5119  0.6075 -0.5297
#> -0.1987 -0.9514 -0.4657 -0.5970  0.1278 -1.6299 -1.1793 -0.4398 -0.9116  0.6820
#> -0.8203  0.6679  2.3214 -1.4328  0.3976 -0.0199 -0.2413 -1.6018 -0.4074  1.2238
#>  0.2998  0.7418  0.3786  1.6767 -1.2742  0.0957 -0.7103  1.8752  2.5805 -0.0029
#> -1.8703 -0.3134 -0.3327 -0.0785 -0.8294  0.4458  1.0581  0.6950  0.0542  1.6530
#>  0.6296 -1.2802  0.7042  1.1002 -0.7369 -1.4735  0.4675 -1.9157  1.6439  0.6879
#> -1.3832 -0.5332  1.4285  0.0814  1.2459  0.3476  0.2004 -0.4381 -1.2596  1.2346
#> -0.4716  1.3713  0.9170 -1.8052  0.2117  1.4498 -1.6182  0.1503 -0.8364  0.7920
#>  0.9629 -0.4104 -0.6443  1.4993  0.4408 -0.7473  0.7552 -1.1494 -0.5853 -0.5704
#> -0.1707 -0.0769 -0.9652 -0.8063 -1.4457 -0.1559  1.7326 -1.6262 -0.4918 -0.9110
#> -0.7448 -1.1541 -0.5405  0.3145 -1.3319 -0.3299 -1.3492 -1.3301  0.6429 -1.0850
#>  0.0454 -0.7224 -0.3350 -1.1099 -0.9148  0.1733 -2.3099  0.5039  0.5561 -0.5777
#> -0.0883  1.1012  0.4142  1.1118  0.0813  0.0594  1.9940 -0.3604  0.0701 -1.0750
#>  1.1810 -1.5404 -1.0119 -1.4410  1.0262 -0.5057  0.5519 -1.0791  0.0528  0.9495
#> -0.2336 -0.1684 -0.7229 -2.1126 -0.1529 -1.3018 -0.1991  0.8201  1.3898  0.4976
#> -0.8471  0.7332 -1.3604 -0.2553  0.1743  0.9536  0.3190  0.4530 -0.8264  0.2736
#>  1.7659 -0.7773 -0.2974 -1.6288 -1.1785  2.3457  0.0211  0.7682  0.9091  2.1692
#> -0.7197  0.0986 -2.3781  1.5422  0.2051  0.4384  0.1207  0.4619  1.1811  0.6162
#> -0.1287 -0.6980  0.1353 -0.2111  0.9027 -0.0059  1.0421 -1.1572  0.0401 -0.2940
#>  2.0506 -1.3305 -0.7102  0.1657  0.0098  1.5171 -0.5016  0.9603 -0.5212  0.3167
#>  1.1512 -1.8490 -0.0520 -1.2120  0.5045  1.6288 -0.7009  0.9350  0.2320 -1.7208
#>  0.1999  0.9771  0.1331  0.3253 -0.8358 -0.1143  0.6550  1.2135 -0.4623  1.0742
#> -1.2092 -0.6650  0.0420 -1.4385  0.5350  0.0738  0.0536 -1.8245  1.2920  1.3688
#>  0.8499 -1.8208 -1.1362 -0.8250 -1.1429  0.0440  0.5493 -0.4742  0.6820  0.4193
#> -0.3603 -0.9430  0.8992 -0.0509 -0.9083  0.1028  1.3267  0.6831  2.5052 -2.1364
#> -0.9463 -1.5920  1.1172 -0.1134 -0.8731 -0.4582 -0.6003 -0.6589  0.5997  0.1357
#> -0.0071  1.4079  0.1809  1.0117 -1.2765  0.2774 -0.3424 -0.8523  0.5534 -0.6963
#> -0.5623 -0.0987  1.4473  0.1035  0.8813  0.2395 -1.1842 -0.0470 -1.0790  0.1260
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]