Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6-6.8221e-01 -2.3953e-01  1.6690e+00  5.4708e-01 -2.6549e-02  8.0439e-01
#>  4.4059e-01  7.1048e-01  2.0823e-01  1.8394e+00  5.6366e-03 -1.9979e+00
#> -9.5893e-01 -1.2128e+00  9.1966e-01 -8.1328e-01 -2.6774e-01 -2.0388e+00
#> -3.3708e-01  3.9572e-02  4.5163e-01  1.3652e-02 -1.3905e+00  2.8079e+00
#>  3.2203e-01 -6.8943e-01 -1.6403e+00 -1.1319e+00  1.2041e-01 -3.8150e-01
#>  8.7409e-01 -1.9737e-01 -1.0430e+00 -2.1029e-01 -3.5956e-01 -2.3336e+00
#> -6.3872e-01 -8.9377e-01  1.7662e+00 -3.7976e-01 -1.9220e+00 -6.7007e-01
#>  5.3099e-01 -2.2494e+00  1.2783e-01  1.4249e+00 -1.3253e+00 -6.9480e-01
#> -1.4029e+00 -1.3175e+00  1.7418e-01  5.8607e-01  1.5131e+00  3.6291e-02
#>  1.3756e+00  4.6355e-02  5.2683e-01  2.0740e+00  2.3578e+00 -5.7311e-01
#> -6.0347e-01  9.3507e-02 -1.3805e-01  1.1027e+00 -1.7093e+00 -1.1026e-01
#>  2.3980e-01 -3.0179e-04 -2.7320e-01 -1.2708e+00  9.5711e-02 -3.8443e-01
#>  3.3646e-01 -4.8504e-01  4.0505e-02  1.1873e-01  1.8915e-01 -2.6453e-01
#>  8.5442e-01  5.8299e-01  1.1406e+00 -5.6384e-01 -3.8093e-01  5.9255e-02
#> -1.1688e-01  8.4395e-01  1.3431e+00  8.4766e-01  1.0309e+00 -3.7845e-01
#>  1.2483e+00  6.7688e-01 -5.9494e-01 -3.2726e+00  4.0394e-01  9.6450e-01
#>  1.4315e-01 -6.3737e-01  2.6881e-01  1.0322e+00  8.6579e-01  2.9935e-01
#> -1.6411e-01  1.7342e+00  1.0070e+00 -9.3840e-01 -1.8756e+00 -6.8867e-01
#>  5.6804e-01 -3.5240e-01 -2.4805e-01  1.6598e-01 -1.7964e-01 -8.9099e-01
#> -1.3745e+00 -1.1183e-01  6.9293e-01  8.2083e-01  2.3479e+00  2.4849e-01
#>  4.1595e-01 -9.2570e-01 -3.8649e-01 -9.1352e-01  2.3017e-01  5.7096e-01
#> -3.3288e-01 -2.1278e+00  1.2090e+00 -2.7441e-01 -2.4131e-02  1.6798e-01
#> -4.1851e-01 -1.1716e+00  8.2884e-01 -8.7459e-02 -2.2643e-01 -1.3917e+00
#>  1.1107e+00  2.2989e+00  8.8012e-01  5.2113e-01 -1.7123e+00 -5.6624e-01
#>  2.1612e-01  2.1220e+00  5.4953e-02  1.0158e-01 -1.1464e+00 -1.4894e+00
#> -2.9485e-01  5.5870e-01  3.0188e-01  9.7828e-01  7.8740e-01  6.1298e-01
#> -5.1494e-01 -1.2320e-01 -6.5364e-01  6.9665e-02  3.9808e-02 -1.6215e-01
#>  2.7424e+00  9.7195e-01 -8.6581e-01 -4.1751e-01  1.3473e+00  7.0009e-01
#>  1.3011e+00 -7.4775e-01  2.1780e-01 -8.6712e-01 -1.4143e+00 -4.7282e-01
#>  9.2201e-02  1.2788e-01  8.7900e-01 -2.7574e-01  8.2052e-01 -5.1312e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]