Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.4242 -1.6965 -0.9739 -0.1249 -2.1359 -1.3445 -1.3425 -0.7162  0.7295 -0.4885
#>  1.7346  1.4043 -0.0207 -0.6590 -0.1138  0.7200 -1.5766 -0.9319  0.1607 -2.2363
#> -0.7350  1.0948  1.0083 -0.9390  1.2119 -1.2168 -1.3590 -0.3436 -0.2844 -0.4357
#> -1.1340 -1.6791 -1.0777  1.7405 -0.2890  2.0974 -0.3985 -0.2973 -1.8096 -1.2873
#>  0.3481  0.0657 -0.0103 -0.3822 -0.4351  2.5813 -0.8400  0.6211 -1.9512  1.7263
#>  0.6994 -0.3490 -1.2150 -1.6968  0.9681 -0.8698 -0.2442  0.3906  0.3589 -0.4108
#> -1.1702 -0.6060 -0.9280  1.5128  0.8644  0.3562 -0.5320  0.5631  1.3897 -0.0259
#>  0.0382  0.5336  0.0472 -0.7868 -0.1079  0.1049 -0.4370 -0.9082  0.0466 -0.4600
#>  0.6201 -0.4027 -0.1253 -1.1483  1.5749  0.3927  0.0144  0.1871 -1.3272 -0.8417
#> -0.3256  0.4065  2.1860  0.5505 -0.3467 -1.1415 -0.2339 -1.3033  0.4309  0.4210
#> -0.0567 -1.0217  0.5313 -1.2597  0.8014 -1.0414 -0.9885  0.2983 -0.1949  0.1892
#>  0.9680 -0.0170  1.4074  0.8868  0.8784 -0.3636  0.4903 -0.5739  1.4634  1.5994
#> -0.9404  1.0250  1.1992  0.7772  0.7704 -0.2934 -0.3118  0.1795 -0.2325  0.3352
#> -0.9064 -0.8863 -2.9628 -1.2640  1.6697 -0.2650  1.1104  1.3668 -0.4805  1.2154
#>  0.5761  0.2110  0.6342 -0.0624  0.1973  0.2330  0.0992 -0.9013 -0.4517  0.7237
#>  0.6179 -1.2866 -0.5851  0.8604 -0.0303 -0.9091 -0.5242 -2.5559 -0.3783  2.1595
#>  1.8457  1.7494  0.6675 -0.0040 -0.8801 -1.9719  0.3593 -0.1901 -1.1443  0.9209
#>  0.3534 -2.3176  0.2810  0.2619  0.9777 -0.5263 -0.4219  1.1766 -0.3553 -0.8266
#> -0.1911 -0.4508  0.9354 -0.0596 -0.9020  0.1937  0.8885 -0.2005  1.2396  1.0022
#> -0.7733  0.1980  0.5248  1.8978 -0.5755  1.0425  0.0893  0.5076  0.5255  0.3600
#> -0.6481 -0.2519 -0.7319  0.1695 -1.8978  1.0368  1.1291  0.6787 -0.2274  1.3855
#> -0.3330  1.0759  1.3273 -0.1375 -1.0984  0.8598  0.2679  1.3876 -0.2491  0.1167
#>  1.6004  0.9863 -1.1782  1.4342 -1.0979  1.0070 -0.8493 -0.3743  0.3864  0.8826
#> -0.8397 -0.4046 -0.5060 -0.6608  0.7062 -1.0085 -0.2164 -1.1821 -0.7184 -0.7877
#>  1.3675  0.8300  0.0536 -0.8764 -0.8694 -0.1672  1.1662  0.2496  0.3492  0.5286
#> -1.2332 -0.1052 -0.0661  1.2196 -1.3324 -0.5018 -0.8602  0.0997 -0.5451  0.6000
#>  1.8629  0.9904  1.0733 -1.0030 -0.4965  0.8700 -0.1587 -0.5879 -0.7931  0.5553
#>  0.5292 -0.4755  0.6959  0.5702  0.3474  0.9516 -0.5448 -1.5753 -0.1360  2.0754
#> -0.9035  1.2637  1.2970  2.2012  0.3381  0.6915 -1.4995 -1.1242  0.8421  1.1832
#> -1.4630  1.1867 -1.3866 -1.1893  2.5255 -0.9390 -0.3619  0.8670  0.4973 -1.5354
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]