Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3087 -0.6318 -0.3822 -0.6803  0.1723 -0.5635  0.0665 -0.3647 -0.2708 -0.8004
#>  0.0968 -1.5924  1.6250 -0.1917  0.3686  1.5207  0.0150  0.3420  0.9593  1.2165
#> -0.3865 -0.6598  0.7606 -0.2411 -0.2478  1.9186 -1.3656  0.6058 -0.4827 -0.9413
#> -0.0400  0.0115  0.5399 -0.9016 -1.1422  0.0032  1.0051 -0.7230 -0.7494  0.8157
#>  0.7915  0.0226 -1.5705 -0.2947  0.3316  0.9046 -0.5544 -0.2428 -1.3180 -1.4522
#>  0.8930 -0.6614  1.7350  1.1790 -0.1249 -0.9472 -0.8527 -1.1001 -1.2373 -0.3105
#> -1.1593  0.8035 -0.4876 -0.4809 -1.5673  0.4484 -0.3265 -0.6869 -0.8779  1.6136
#> -0.7425 -1.0417  0.3640 -1.2339 -0.5632 -0.8195  1.1682 -0.3829 -0.1209  1.7312
#> -0.0149  0.0940  0.8410  0.6403 -0.3966 -0.3789  0.0428 -0.0389 -0.2916  1.8298
#>  0.4391 -0.4973  0.0311  0.6268  0.5393  1.3610  0.8376  1.2904 -0.7412  1.3622
#>  0.5476  0.3644 -0.7475  1.0359 -2.5451 -0.2621 -0.6599 -0.8815 -1.1324 -0.7122
#> -1.2618 -0.0587 -1.5926  0.9579  1.0182  1.0090  0.8420 -1.2416 -1.0889  0.9790
#>  0.0257 -0.5972 -1.5642  0.4999  0.7514  1.5946  0.2887  1.1183  0.6669  1.3188
#> -0.0275 -0.4950  0.4812 -1.5471 -0.2066  0.8958 -1.8734  2.9056  2.5137 -0.2773
#> -0.4648  0.2050  0.1936 -1.5800 -0.4605 -0.5645  0.1025  1.6379 -1.0378  0.1286
#>  0.6730 -0.0061 -1.1180 -0.1514  0.1059 -0.4972 -0.6228 -0.0436 -1.9462  1.8151
#> -0.5435 -0.6488  0.2915 -2.4311  0.4627 -1.9468  1.2427 -0.3569  0.0085 -0.6766
#> -0.2653  1.1940 -0.0093 -1.0104 -0.1758  0.8107 -1.0512  3.0154 -1.1362  0.5890
#> -2.0822 -0.0012 -0.4808 -0.3559 -2.1406 -1.1323  0.7960 -1.0146 -1.8761  0.6323
#> -0.7344  1.0029 -0.4858  1.2241 -1.8342  0.7662  0.2898  0.0913 -0.1405 -1.2083
#>  0.0899  1.9035  0.4684  0.9734  0.5527 -0.1429 -2.3318  2.0454 -1.3306  0.2125
#>  0.8875  0.2776  0.5984 -0.3308 -0.8860  0.5510 -0.5997  0.5207 -0.2075  0.0351
#>  0.9268 -0.6757  0.6379  0.6671  0.3975 -0.1779 -0.0695  0.6049 -1.0694  0.2905
#> -0.3161  1.6733 -0.2733 -1.3163  0.0664  0.9534  0.0555  1.1363 -0.2317 -0.5455
#> -0.2960  0.3994 -1.1353 -0.5319  1.2429  0.0139 -0.8116 -0.9502  0.6735  0.4988
#>  0.6354 -0.1675 -0.1413 -0.1363 -0.9091  0.8073 -1.1345 -0.7461 -0.7900 -0.2710
#> -1.0585  0.3319  1.2266  0.2379  0.4816 -0.5669  1.3120 -0.2563 -1.2215  0.0747
#>  0.3913 -0.9942  1.1746 -0.7549  0.3957 -0.3301 -0.3411 -0.1246  0.7752  1.7734
#>  1.4564  0.8629 -0.3658 -0.8307  1.7268  0.5311  1.3230  0.7140 -1.1843 -0.3663
#> -0.0597  2.2006  1.0221  0.5514 -0.7330  1.4953 -0.3042 -0.1144  0.3554 -1.5904
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]