Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.5076  0.4344  0.4635 -0.8208  0.1298 -0.2269 -1.5850  1.7039 -0.4505 -0.4430
#> -0.1168 -0.2329 -1.5512 -1.2266 -0.0981 -1.0097  1.2861 -0.1380  0.8283  1.4965
#>  1.3552 -0.5138 -1.0746 -1.1320 -1.3838  0.0807 -0.0138 -0.9696  0.3121 -0.0733
#> -0.4676 -0.2295  1.8774 -0.6158  0.0184  0.0702 -0.6397 -0.3983 -0.9034 -0.8609
#> -0.3024  0.6400 -0.3453  2.2294  0.4801 -2.3559 -0.5846  1.6212 -0.6144  0.6596
#>  0.5594  0.3987 -0.5258 -0.1351  2.0886 -1.0602 -0.2604 -0.5428 -0.4182  0.0642
#> -0.9732 -0.2771 -0.7477  0.4838  0.7973 -0.4716  0.2052 -0.6287  0.3277 -0.3503
#>  1.4073 -1.4971 -0.7753  1.8124 -0.3205 -0.0580  0.0704 -0.9930 -1.7545  0.5103
#> -1.1274 -1.5196  0.4455 -0.2707  1.1986 -0.1012  0.6295  0.6431 -0.0146 -0.7147
#> -1.7577  0.3505 -0.8906 -0.1035  0.4239 -0.6597  0.3860 -0.1984  0.0612  0.2105
#> -0.0938  0.0982  0.1077  0.2975  1.0648 -2.9301 -0.4784 -0.4957 -0.6138  0.6882
#>  0.0834 -1.9713 -1.7214 -1.6089 -0.3026 -1.4688 -0.9532 -1.1226 -0.1417 -0.1598
#> -0.0114 -0.7388  0.2073 -1.4376  0.7684 -1.1105 -0.9786  0.7405 -0.0939 -0.2311
#> -0.2795 -0.3441 -1.4994  0.7816 -0.9140  0.6826  1.6380  0.3351 -0.1655 -0.6857
#>  0.1308  0.0418  0.0633  1.7966  0.3981  0.0229 -0.5492  0.7939 -0.7779  2.1274
#> -0.1367 -0.1440  0.8869  0.4752  0.3545  0.5104 -0.2370 -0.0950  0.5958 -1.2297
#>  0.2847 -0.2208 -0.9520  1.1964 -0.3178  0.3061 -0.0515  0.3396 -1.7642 -0.2048
#>  0.6334 -0.7804 -1.0035  1.4565  0.0567 -0.7895  1.5936 -1.0575  0.4353  0.6864
#> -0.1726  2.4618 -1.6975  1.3034  1.3215 -1.2671  1.4500 -0.2302  0.6688  1.1314
#> -1.2123  2.0861  0.3114  1.0447  0.1144  0.9596  2.2103  1.4696  0.0681 -1.8476
#> -1.5145 -0.4795 -0.4595 -1.2507  2.4260 -0.2481  0.1017  0.2725 -1.3378 -0.3808
#>  1.2562  2.2463  1.9659  0.9371 -0.5247  0.1172  0.5425  0.4101  0.1343 -0.3843
#> -0.9460  1.0440 -0.8182 -0.4441 -2.0939 -0.0179  1.5196  0.6398 -1.6917  0.2491
#> -0.0771 -1.7474  0.1691 -0.3644  1.4497 -0.9964  0.3722 -0.0556  0.3298 -1.0904
#> -1.0868  0.1040 -0.4870  0.0220 -0.9897  0.4337  0.1909  0.0396  0.2840 -0.9345
#> -0.3045  0.4152 -0.7904 -0.2304 -0.4182  0.8097  0.5511 -1.5241  0.9439  0.3277
#>  0.4743  0.5040 -0.3620  0.5445  2.1737 -0.2994  0.2342  0.1785 -0.2997 -1.1727
#> -0.3741  2.0667 -0.2600  0.3059  0.0266 -0.8382 -0.6993  1.6568  0.4835 -1.2820
#> -2.4021  0.8385 -1.1353 -0.1974  0.6067 -2.4839  1.1948  0.0868  0.9683 -1.4667
#>  0.1789  1.1765 -1.3408 -0.1201  0.2804 -0.9522  0.2189  0.0877 -1.0153  2.0893
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]