Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.1389 -0.0320  0.5003  0.1911 -0.1160  2.9799 -0.4857 -0.9575 -0.2896  0.8337
#> -0.9151 -0.3031  0.3956  0.0003  0.1513 -0.0093 -0.8517 -0.2757 -1.8053  0.2504
#>  1.2427  0.0117 -0.4600  1.5559 -0.0690 -0.1365 -0.2635 -2.4620 -1.0138 -0.4416
#>  0.9743  0.2665  0.6548 -0.2882  0.5886 -0.9336 -0.4675  0.2238  0.8526  0.2335
#> -0.6970 -0.9965 -1.1396 -1.0368  0.2005  1.3930  0.2303  0.0129  0.0256 -0.6214
#>  0.1289  1.8733 -1.7843 -0.4275 -1.6528  1.4987 -0.2475 -0.9571 -1.8853 -0.9421
#> -1.9627  0.6110 -0.3309  0.7817 -0.9264  0.6446  0.3519 -1.9111  0.4678  1.4374
#>  1.2945 -0.8128 -0.5847  0.5679 -1.4401 -0.9520 -0.5656 -0.8048  1.6989 -2.0563
#> -1.1162  0.9654 -1.5862 -0.9908 -0.8658  0.1837  0.5175  0.9040 -0.1394 -0.2272
#>  0.7344  0.5830  0.2540 -2.1854 -1.0355 -0.3990 -0.0630 -2.3986 -2.8417 -1.1227
#>  0.2839  0.1959 -1.2376 -1.0517  1.2729 -2.5468  1.2849  0.2730 -0.3354 -1.0924
#> -0.2866  0.5604 -0.2930  1.4926  0.4077 -1.0043 -1.2778 -1.7315  0.3281 -0.9626
#> -0.3803  0.4320 -1.4691 -0.2044  0.4474  1.4969  0.9851 -0.8574 -1.4009  2.7154
#> -0.7048 -0.8163 -0.5836 -0.5779 -0.7686 -1.2699 -1.1435  1.0373  0.7516  0.9608
#>  1.1521  1.1631  0.9600  2.1691 -0.8209 -0.7631  1.2873  0.1597 -1.2239 -0.0940
#> -0.0222 -0.4381  0.8531  1.8384 -1.2132 -1.3387  2.2247  0.6444 -1.3745 -0.4953
#>  1.8910 -0.0605  2.0306 -1.4943 -1.2274  0.0457  0.0195 -0.3604 -0.0351  0.8680
#> -0.2084 -0.4859 -0.8351  0.7378  0.4795 -0.5458 -1.1516  0.3606 -0.6206 -0.9062
#>  0.5521  1.5816  0.5840 -0.4105 -0.8477 -0.3720  0.9630  0.3961  1.9677 -0.5026
#>  0.2719  0.0400 -0.7801 -0.3166 -0.3674  0.3202 -0.1658  0.3823 -0.6916 -0.1542
#> -0.9669  0.1363 -1.6257 -0.9749  0.3867  0.9985 -0.8106 -0.5995  0.2430  0.5784
#> -0.2141  0.5457 -0.6719 -1.9233  0.3750 -2.6305 -0.1393 -0.1998 -1.4253  1.7943
#>  0.2903 -0.7915  0.6735  1.8261 -0.6823 -0.5619 -2.1479  0.6482 -1.3710 -0.4740
#>  0.0357 -0.0392  0.6055  1.1731  0.4000 -0.6975 -2.8157 -0.7078 -0.3420 -0.9088
#>  0.0765  0.7013 -1.2579 -2.0838 -1.4017  0.7497 -1.7421  0.8615  1.9864  1.6321
#> -2.7710 -1.5998  1.5379 -1.3452  1.2715 -0.2980 -0.7518  0.5239 -0.4393 -0.3541
#> -1.2369  0.0466  0.4707  1.1651 -1.4077  0.9040  0.7627 -0.9600 -1.0140  1.0625
#>  0.4506  1.2998  0.1896 -0.4607  0.8592  1.2450 -0.5039  0.4510 -0.1661  0.8991
#> -0.5175 -0.4349  1.7618  0.0985  0.6433 -1.5110  1.7968  0.0195 -0.5706 -0.3530
#> -0.8771  0.8454 -0.2214 -0.9590  0.7813  0.2297  0.9000 -2.1256  0.5656  1.5268
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]