Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.8477  0.5008 -1.6762  0.8975 -1.4536  0.3289 -1.2094 -0.2142 -1.1207 -1.1379
#> -0.4939  1.8229 -1.7990  0.5475 -0.7551 -0.5565 -1.2466 -2.3254  0.5293  2.4853
#>  0.0350  2.0810  1.0017 -0.9641  0.4918  0.8762 -1.5228 -1.7740 -1.3198 -0.3518
#>  0.7146  1.1022 -0.9037  0.3336 -2.1605  0.2760  1.7868  0.7310 -1.4955  0.0388
#> -0.7648 -0.1337 -0.1852 -0.3376  0.3507 -2.7668 -1.5619 -0.8367 -0.7835 -0.1148
#>  1.0054  0.6986  1.3300 -0.1241  0.8580  0.4632 -0.1540 -1.6059 -0.5178 -1.3290
#> -1.4047  0.0847 -0.0119  0.0638  1.4144 -0.3487  0.0241  0.1766 -1.1089 -0.1615
#> -0.8446  0.8330  1.7423 -1.6412  1.1350  0.5893 -0.7370  0.1657 -1.8570 -1.1808
#>  1.2098 -1.4943  0.3276  1.4450 -0.6291  2.8153 -0.2865 -1.7047  0.4541 -1.4490
#> -0.3525  0.5369 -1.5771  0.9808  0.9945  0.0762  0.2638  1.0657 -0.5307  0.2225
#>  0.1465 -0.3982  1.5035  0.0416 -0.4980  0.6262 -0.6501  1.0657  0.0276  0.0897
#>  0.0077  0.5089  0.4363  1.8834  0.6415  0.5691 -0.2446  1.5418  1.6339  0.5010
#>  0.8540 -1.1897  0.7807  0.8887 -1.8803 -2.0280 -0.0427 -2.8332 -1.2373 -1.1621
#>  0.8961 -0.9139 -1.4624 -1.3934  0.1721 -0.0118  1.5097  0.1504 -0.3531  0.6478
#> -0.3150 -1.4192 -0.0655  0.6186 -0.6120  1.5896  0.7312  0.7405 -1.2727  0.1927
#>  0.3901  0.5584 -1.4913 -0.5808 -1.7389 -0.4696  1.1767 -2.1304  1.1185  0.0230
#> -1.0889 -2.9744 -0.1040 -1.1090 -0.2463 -0.3329  0.2054 -0.5773 -1.3694 -1.0903
#> -1.3885  0.3765 -1.5848  1.9894 -0.7089  1.3172 -0.0496  1.0790  0.2016  0.5770
#> -0.3297 -0.2635  0.7495 -0.5410 -0.0136 -1.4289 -1.2696 -0.7066  1.5955 -0.0646
#>  0.9029  0.6847 -1.9193 -0.3772  2.0133 -0.1698 -1.2491 -0.8940 -0.7276 -1.3995
#> -0.3342 -0.7666  0.9111  0.8853 -0.6454 -0.4249 -1.3427 -0.4140 -1.6757 -1.0352
#> -0.5458  1.8651  0.4299 -2.3064 -0.8940  0.8148 -1.0913  0.1494 -0.0276 -0.9255
#> -0.8736  0.6118 -0.9207  0.8739 -0.6702  0.5402 -0.3686 -0.1554 -0.8144  0.7603
#>  0.9125  0.4366 -0.3711  0.2116  2.7587  0.6775 -0.3572  0.0279  0.3740  0.1976
#> -1.2122  0.5984 -1.6291 -0.6897 -0.3879  0.8017 -1.5089 -0.1030  0.0239 -0.2527
#>  1.0937 -0.8048 -0.4900 -0.9741  1.5165 -1.3893  1.7715  0.1727 -1.2304  0.4376
#>  0.0557 -0.2833  0.6788  0.0628  0.9637 -1.0145 -0.8725 -1.4743  0.3987  0.4724
#>  0.5368  0.5250  1.2774  0.6614  0.6121 -0.9695  0.9068 -0.5869 -2.5515  0.2764
#>  0.2123 -1.0751 -1.1722 -0.8761  1.9801  1.3909  0.3093 -1.9591 -1.7065  1.0247
#>  1.0031  0.1291  0.1172 -1.3193 -0.3936  0.3818 -0.4334  0.2531  1.6396  0.8107
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]