Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.7177  0.9350 -0.5794  1.1625  0.4913  0.3179  0.3702 -2.7310 -0.7807  0.8948
#> -1.8638  1.2673  0.6425  0.2183 -0.1960  0.8535  1.2066  0.0557 -2.2152 -1.1901
#>  3.0666  1.4892  0.6097 -0.5308  1.0018  0.4923 -0.1891  0.0147  1.6579  0.9706
#> -0.9963 -0.4061  2.2610 -1.5621 -0.2083  1.3228 -0.4098 -0.4088 -1.3253 -0.4523
#> -0.5761 -0.5781  2.2098  2.3003  0.3793  0.6936 -1.5398  0.7607  1.0822 -1.4740
#> -0.9341  0.3702  1.2060 -1.3330  1.5628 -0.8957 -0.1990 -0.3211  1.5855 -1.8419
#> -1.2014 -0.3948 -0.9399  0.8541 -0.2937  1.0688 -0.6379 -0.3351  1.3912 -0.7168
#> -0.2423  0.7815 -0.5114 -1.3029 -1.0623 -1.5160  0.8789  0.9998 -1.2374  0.6711
#> -0.6529  1.4975 -0.3655 -1.2891  1.9027 -1.9381  1.0465  0.8721  0.2298 -0.1961
#>  0.2023 -0.5937 -0.1006  0.0032  0.5894 -1.8409  0.8018  0.0970  0.4228 -0.2540
#>  0.2701  0.2083 -1.0592 -0.3878 -0.2608  0.4173  0.2048 -0.5190  0.2965 -0.0403
#>  0.1216 -0.6646  0.1433  0.3271 -0.4890  0.5212  1.4108 -0.7148  0.4441 -0.3567
#> -0.1619  1.8282  0.0138 -0.3219 -0.8882 -0.2465  0.6359 -1.2286 -0.6557  1.6530
#> -0.2202 -0.2857 -0.3306  2.1516 -0.3527  0.2755  0.6048  0.0129  1.9593  0.1586
#> -1.1030 -0.4558  0.7759 -1.0968  0.7409  0.4826 -0.4890  0.4905 -0.2613 -0.6642
#> -0.0305 -1.4536 -0.5389  0.9193 -0.9978 -0.9553 -1.0923  0.2750 -0.4828 -0.0393
#> -2.1052 -0.6884 -1.5053 -0.6049  1.7446 -1.2091 -2.1595  0.8194 -0.2672  1.1178
#>  0.8846 -0.0379 -0.6043 -0.3544  0.5971 -0.2322 -0.9552  1.2353 -0.0444 -0.6027
#> -1.5518  0.5159  0.2956  0.2820 -0.0887 -0.0963 -1.5740 -1.0702 -0.5833 -0.4238
#> -0.6459 -1.1857 -0.5540 -0.7218 -2.3430  1.2128  1.0061 -0.8910  0.1319 -0.2735
#> -0.0422 -0.6413  0.8092 -0.4023  0.5984 -0.0899 -2.6215 -1.2711  0.2736 -0.1303
#> -1.0845  0.3954 -1.1461  1.3281  0.6125 -0.1472 -0.7755  1.6715  0.1913 -0.6351
#> -0.3002 -0.8077  0.0277 -0.1138 -0.0482  0.7451  0.9908  1.0000 -2.4063  0.1705
#>  0.4990 -1.1877 -0.7320  1.6989  1.5403 -1.3873 -0.1543 -1.2274 -2.9361 -1.8622
#> -0.2690  1.3501 -1.6150 -0.8682  2.0278 -1.6083  1.8386 -1.4285 -0.4711 -0.9643
#>  0.6955  0.2950 -0.9320 -0.1650 -1.4048  1.5456 -1.1370  2.1490  0.4412 -0.1506
#> -0.0640  0.4601  0.0053 -2.4817 -1.3690 -0.7148 -2.0610 -0.4574 -0.8057 -1.1058
#>  0.1286 -0.0749  0.5500  1.8450 -0.1321  0.9209  0.9088  0.0014 -0.0584 -0.2058
#> -0.0967  0.2198 -0.0441 -0.8993 -0.7636 -0.0594  2.3159  0.5474 -1.0761 -0.2935
#> -0.6442 -1.8340 -1.0142  0.1135  1.1170 -0.4034 -1.4242 -0.8026 -0.9368 -0.9430
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]