Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5015 -0.5332  1.0255 -0.3876 -0.3370 -0.0982 -0.8208  0.0262 -0.2940  0.0204
#>  0.3031  0.0383  0.2440 -0.1570  1.2856 -0.4142 -1.8909 -1.2235  0.6269 -0.1242
#> -1.0424 -0.1391 -1.2165 -0.1418  0.7932 -1.3174 -0.1156  0.2586 -1.3326  1.2944
#>  1.6113 -0.3391 -0.0111 -1.5023  1.0007 -0.6599 -1.9648 -1.0707  0.6523  1.5779
#>  1.9089  0.5618  0.0990  0.2515 -1.9838  0.0322  1.8113 -0.3209  2.1357  0.8734
#>  0.4783  1.1104  0.3905  0.2363 -1.6406 -0.8289 -0.5398  0.8408  0.3678 -0.0023
#>  1.2783 -0.1415  1.1925 -0.3593 -1.0917  0.4187 -0.0374 -1.2894 -0.3092  0.7586
#>  1.7025 -1.0373  0.7930  2.0390 -2.3183 -0.7213  0.9007  0.0776  0.5519 -1.1461
#>  1.1502 -1.3852  0.1387  0.7534  0.5581  0.5168 -0.7679  0.1581 -0.7775 -0.9303
#> -0.8942 -0.2708  0.7593  0.7934  1.4125  1.2609  1.3173  1.4173 -0.3999 -0.0420
#> -0.6317 -0.1769  0.3636 -0.8125  0.6839  0.9323  0.2323  1.0069 -0.4803 -0.5219
#>  0.4282  0.1771  0.5367  1.2910 -1.4336 -0.6622 -0.4407 -0.6695  0.1117 -0.7843
#> -1.7675  0.1854 -0.3497 -0.2600  1.3224  0.9801  0.1637 -1.1840  0.7134  1.4757
#>  0.1430 -0.0074  0.4796  0.7898  1.6438 -0.9677 -0.2332  1.0710 -0.2469 -1.0359
#>  0.1912 -1.1767 -0.0083 -0.4943  1.6060 -1.0626 -0.7641 -0.6782 -0.4814  0.1848
#> -1.1173 -0.3938 -1.0444  0.2256 -0.6976  1.5879  1.4958 -0.9155  0.4299  1.0085
#>  0.7710  0.5457 -0.1682  0.9935 -0.2606  1.0909 -0.1651 -1.1578  0.4940 -0.1586
#> -0.4421  0.0874  0.4439  0.9733 -1.7452 -0.2611  0.1282  0.4685 -2.0742  0.3607
#>  1.0606 -0.6639  3.4576 -1.4622  0.6964 -0.0539  0.4562 -0.0776 -1.4925  0.0191
#> -0.1397 -1.8310 -0.0537 -0.1198  0.3331 -1.1564  1.0406  0.0868 -0.8582  0.5037
#> -1.0613  1.7867 -1.1365 -0.0237 -0.5916  0.1091 -0.2839  0.5169  0.6855 -1.2683
#>  2.2876 -0.3416 -1.8034 -0.1897  0.7968 -0.0709  0.0935  0.5057 -1.2276 -0.9557
#> -1.4762 -0.2861 -0.1893  1.4514  0.3693  0.3522  0.6716  0.6790 -0.4087 -1.3589
#> -0.2833 -1.1555  1.0924  0.5950  1.6586 -2.0505 -0.4088 -0.5393 -1.0513 -0.6505
#>  1.8798  0.6928 -0.5962 -0.1416  0.8803  1.0577 -0.4639 -2.4350  1.6907 -0.0108
#>  0.8088 -0.3720 -1.0317 -1.0256 -0.2073 -0.1611  2.8337 -2.3601 -0.1602 -1.7121
#> -0.3018 -0.5746 -0.9192  1.3105  1.2632  0.3401 -1.1782  0.3613 -0.6875  1.5326
#> -0.7680 -0.0957 -1.4331 -0.4699 -0.2173 -2.3628 -0.1222 -0.2976 -0.4058  0.1375
#> -1.0024  0.0116 -0.7849  0.9846 -0.9093  0.2597  0.0780 -0.6827 -0.8802 -1.2967
#> -0.8440 -1.8800  0.3854 -0.1816  0.5841  1.0849  1.2741 -0.4393 -1.5160 -1.3645
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]