Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.6940 -0.6918  0.2915 -0.0412 -0.2663  0.1941  1.0969 -0.1975  1.9009 -1.7082
#> -0.6198 -0.0344 -0.1796 -1.6578 -0.8735  0.9538 -1.5127 -0.4950 -0.2353  0.7920
#>  0.0471  0.3601  1.1714 -0.4657  0.6988 -0.6876 -0.3962 -0.0496  1.3318 -0.2803
#>  0.4226  1.4043 -0.5123 -0.9709 -0.5980 -1.6445  0.0260 -0.3364  1.4429 -0.4215
#> -0.1071  1.2577  1.0421 -0.4876 -0.8570 -3.0289  0.0089 -0.2366  0.7482 -0.5344
#> -1.0325 -1.2652  0.1819  0.3605  1.3671  0.4251  0.0740 -0.5216 -0.3594 -3.0906
#> -1.3699  0.7488  0.5262  1.0753 -0.2207  0.3221  0.0079  0.1948 -0.2177  0.7018
#> -0.7323 -1.2942  0.4058  0.5871  0.0395  1.2505  0.2854  0.2007 -0.8594  0.6944
#> -0.5348  1.0836 -0.1638  0.9532  0.7229  0.4709 -0.3436 -0.8410  0.6314  0.0071
#>  1.3625  0.5857  1.8947  0.8418  1.7487  0.0671  0.5817  0.1054 -1.7925  0.7938
#> -1.8198  0.3442 -1.6379  2.5477  1.7046  0.3150  1.0566 -0.9332 -1.8090  0.7337
#> -2.4100  0.9371  1.1044 -0.2071 -0.9486  1.0686 -0.5369 -0.3789 -1.1140 -0.2919
#>  1.2276  0.1890  0.0100 -0.4334 -3.3676  1.3061 -0.1162  0.1501  0.4135  1.5371
#> -0.2764  1.7901  0.0314 -1.5046 -0.8090 -0.2967  0.3798  0.0379 -0.6341 -1.0771
#> -0.1622  1.3302  0.4193  2.1177 -1.2131  0.2116 -0.3516 -1.6103  0.7222 -1.2991
#> -0.7770  0.0650  1.1266 -2.4759  1.7042 -0.4265  0.4360  0.0871 -1.2230  0.8707
#> -0.2818 -1.6759  0.9713  1.2207  1.3559  0.9131 -0.2499  1.3127 -0.0637  0.1300
#>  1.7011 -1.2932  2.4201  0.1735 -0.7963  1.6091 -1.3523 -0.4298  0.0612  0.8774
#> -0.4517  1.0215  0.4462  0.1995  2.1885 -0.2509 -0.3281 -1.4138 -0.2218 -0.4184
#>  0.8996 -0.6656 -0.9831 -0.5017  1.0537 -0.4076 -1.1225  1.3194  0.4359 -1.1122
#> -0.0669  1.8748  0.4741  0.4343  0.3441  1.0998  0.0939 -0.1872  0.2405 -1.4975
#>  0.9596  0.4266  2.0108  0.3590 -0.9535  0.1342 -0.3807 -2.1756  0.3641  0.3200
#> -0.4405 -1.7630 -0.7211  0.2377 -0.8433  1.8597  0.9333 -1.4728  0.0671  0.2232
#> -0.7628 -0.5926 -0.6907 -2.2222 -0.3374 -0.0343  0.6869  1.0121 -0.4810 -0.5490
#>  0.4019 -0.2403 -0.9424 -0.8720 -0.6614 -1.0901  0.1224  0.0178  0.5977  1.0970
#>  0.6804  1.7974  0.2234  1.8037  1.6569 -2.1750 -1.0793  1.5264 -0.0858 -0.4349
#> -0.2265  0.9321 -0.8310 -0.1576 -0.2082  0.6701  0.2819  1.0744  0.7929 -0.5526
#> -0.8443  0.6955  0.1450  0.6715  0.3168  0.5357 -1.1014 -0.6696 -0.0133  0.0580
#>  0.1177  0.2962 -0.6882 -0.1212 -0.3607 -1.5441  0.6280 -2.0143  2.1460  0.4253
#> -0.9328  0.3628 -0.9214 -0.9909  0.3281 -1.7163  0.1023  0.8026 -0.5141 -0.4301
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]