Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-2.2683 -0.7786 -0.6993  0.5745  2.7545 -0.1959 -1.7115 -0.3476  0.9807 -1.0430
#> -0.1773 -1.2149  0.0771 -0.5114 -0.1596 -1.1001  2.4144  0.5345 -0.5422  1.1328
#>  0.1474  0.3715 -0.8742  1.3900  0.3829  1.3655 -0.8875  1.4059  0.2645  0.0874
#> -0.0594 -0.5952 -0.5735 -0.0746  1.0624  0.1503 -1.7500 -0.5408  1.7429  0.6852
#> -0.4401  0.2828 -0.0897 -0.2163  0.2942 -1.0662 -0.4275 -0.0597 -0.7620  1.6628
#>  0.3147  1.4112  0.3547  0.1470  1.4272 -0.0385 -0.8352 -1.4276 -1.2206  1.0930
#>  0.8504 -0.8207 -0.5817  0.0643  0.5516 -0.8367 -0.0017 -0.7685 -0.0327 -0.2740
#>  0.1788  0.9011 -2.3576 -0.6999  1.1015 -0.1378  0.9216  0.5571 -0.8196 -0.7977
#> -0.4700 -1.0541  0.4614 -0.7380 -0.9736  0.6733  0.6513  1.0247  0.8176 -1.1267
#> -0.7890  1.1735  2.5915 -0.4384 -1.2554 -0.1254  0.0793  2.0619  0.2833  0.2915
#> -1.5002  0.2305  1.0254  0.8489 -0.9024  0.6133  1.0719  0.6716 -0.4594 -0.4021
#>  0.4989 -0.3088  0.8783 -1.0705  0.2494  0.0582  0.2921  0.8094  1.2456  0.9905
#>  0.0634 -0.2955  1.5178 -0.1639 -0.2777 -0.3467  1.3951  0.0493  0.0143  0.3334
#>  2.7272  0.2013 -2.2016 -1.7640 -0.0311 -1.8361  0.0759  1.3691 -1.0271 -1.8296
#>  0.9509  0.6427  0.9181  0.4937 -0.7711 -0.8353 -0.1490  0.7861  0.3040 -1.2040
#>  0.3091  0.1747 -0.6377  0.6138 -0.4302 -0.9420  0.5923 -0.3249 -1.5987  0.2972
#> -0.0211 -0.3946  0.1093  0.1605  0.3878 -0.6951  0.6886 -0.1559  0.0646  0.4195
#> -0.8364 -0.6377 -0.8979  1.5057 -0.4935  0.3440 -0.3774 -0.2712 -0.4418  1.2146
#>  0.1641  0.2648  2.3670  0.7294  1.3635 -0.0834  1.5817 -0.0353 -0.4824  0.6572
#>  0.0832  0.3792  0.8441  1.2485 -0.0130 -1.5668  1.2579  0.5863  0.8261 -1.4468
#>  1.2641  1.9847 -2.7498 -1.3150  0.7936 -0.4552  0.6068  0.9254 -0.0400  1.0624
#>  1.6589  0.4124  1.7153 -0.3751 -2.3495 -0.1319  0.9155 -0.5116  0.0088  0.3126
#> -1.2356  2.6251 -0.4719 -0.9334  0.0441  0.3277  0.1238 -0.4195  0.5561 -0.4364
#> -0.0111 -0.9392  1.1762 -1.9172  2.1643 -1.8548 -0.9159  0.1984  0.0234  0.6163
#> -0.5514  1.1747  0.3544 -0.2068 -0.4162 -0.8317  1.5709  0.2447  0.5826  0.9628
#> -0.3783 -1.0977  0.8314  0.5149  0.1578  0.5469  1.3453  0.0453  1.3752  1.0912
#> -1.7583 -1.3211 -0.9626 -0.4440 -0.5424 -0.3655 -0.9678  2.0205  0.3509 -0.3235
#> -1.9340 -0.3694  0.3607  0.2768  1.2038  0.7441 -0.6485  1.4019 -1.3015  0.9832
#> -0.5143 -1.0134 -0.8636 -0.5802 -1.6159 -2.4272 -0.0911  0.5787 -0.2039  0.1082
#>  0.0503  0.4762  2.8510  0.4530 -0.1949 -0.9076 -0.0979  0.7461  0.6137 -0.2328
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]