Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.9870 -1.5813  1.8453  0.0161  0.5543 -0.6564 -0.4876 -0.7923  0.1461 -0.4220
#> -1.4236  0.5270 -0.2602 -0.6024  0.6781 -0.1015  0.9501  0.0627  0.0631 -1.0615
#> -1.9041  0.9427  2.1845  0.7487  1.1877  0.9342 -1.3162  1.4424 -0.8756  1.5619
#> -1.6731  0.3625 -0.6160  1.9673  1.3640 -0.8315 -0.2472 -0.2886  0.1042 -0.4485
#>  1.3276 -0.6358 -1.6557 -0.5674  0.5838  1.2229  0.4089  0.6182 -0.0461  2.0470
#>  0.7999 -0.4100  0.0490  0.8472  1.4659  1.1422  0.1531 -0.0501  0.3320 -0.9461
#> -0.2500  1.5174 -0.1175 -1.0947  0.3989 -0.5832  2.2282  0.3579  3.0781  0.9715
#> -1.2625 -0.5785  0.3983 -0.3834 -0.6285 -1.1821  0.5251 -0.4236 -2.2062 -0.1162
#> -0.3199  0.0496 -0.5773  1.5359 -0.8284 -0.6269  0.6978  1.3441  0.1295  0.2846
#> -0.6303 -1.3558  1.8457 -0.1486  0.9603 -0.3906  1.4406  1.4865  0.2871  1.0341
#>  1.1235 -1.4705 -0.8770 -0.0689  0.3298  0.1899 -0.7916 -0.0065 -1.0087 -1.1138
#> -0.8914 -1.4215 -0.7425  0.1327  0.1528  0.4291 -0.6709 -0.6761  0.4106  1.5363
#>  0.6178 -0.2978  1.0601  0.2648  0.4535 -0.2263  0.8632  0.9461 -2.2499 -0.4933
#> -1.5387  1.4605 -0.3491  0.5986 -0.5567 -0.4767 -0.1136  1.1454  1.5720  0.0587
#>  0.4942  0.8382  0.3231 -1.3084  0.1711  0.1657 -0.5714 -1.1877 -1.6714  0.2599
#>  0.5716 -1.2347 -0.5892 -1.7833 -1.1001  0.2969 -0.8674 -1.1121  0.0848 -0.5375
#>  0.2703 -1.2057 -1.0140 -0.9335 -0.5658  1.6932  1.2235 -0.5871  1.4473  1.2611
#> -0.3761 -0.4176  1.1416  0.8461 -0.8375  0.1151  1.2555  0.2471  0.5983 -1.2210
#>  0.6999 -0.6130 -2.2244  0.4236 -0.4155  1.0789  0.6935 -1.0927  0.3477 -0.2497
#> -0.6285  1.3681 -0.0496 -0.6772 -0.4740 -0.8362  0.0706 -1.4590 -0.3215 -1.3023
#> -0.0415 -0.8021  1.0323 -0.7776 -0.1021 -0.3228 -0.1966 -0.1829 -0.4321 -0.1080
#>  0.4099 -0.1656  1.2172  0.6750  0.0755  0.9802  0.0695 -1.1366 -0.5642 -0.9789
#>  0.0246 -1.9325  2.3275  1.1393  1.1814  1.6949  0.8286  1.2414  0.8916 -1.2227
#> -0.4067 -2.3779  0.1339  1.3531  0.3575  0.4083 -0.7394 -0.2282  1.2883  0.1470
#>  1.1412 -0.1547  0.3892  0.1512  0.8028  0.3158  1.4472  1.1641 -0.0662 -1.0983
#> -0.1968 -0.3643  0.2469  0.9876 -0.0912  0.5541  0.1629  0.1865 -0.1115 -0.1557
#> -0.5258  0.1677  0.0025 -0.9057 -0.2245 -0.8695 -0.0217 -0.7245  1.5284  1.6936
#>  0.5193 -0.3684 -0.7235 -0.8571  0.8660  0.0581 -0.5877  1.2127 -0.3043  0.3904
#> -1.3822  0.1588 -0.1218  0.4073 -1.2269 -0.6283 -0.6909  0.6590 -0.3351  0.5466
#> -1.5008 -0.2511  0.4752 -0.1956 -0.9999  1.9520  0.3082 -0.3847 -1.1558  0.2661
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]