Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.2541 -1.0349 -0.6070  0.2753 -0.6876 -0.3640 -1.2370  0.6212 -0.1135  1.3434
#>  0.1911 -1.4655 -2.5109  1.5806  0.2216 -1.6302  0.9702  1.0099  0.6140  1.7750
#>  0.0932 -0.4543 -0.8724  0.1307 -0.6375 -0.5061 -0.7213 -0.0598  0.9948 -2.3304
#> -0.5848 -0.7954 -0.5999  0.7537  0.0421  0.2627  0.8391 -1.0680 -0.8597  0.2342
#> -0.9006 -1.0086 -0.6475 -1.3870  1.1745 -0.5579 -0.5079  0.0913 -1.5377  0.1353
#> -1.3252  0.5561  0.9770 -0.3659  0.3198 -1.2170 -0.1635  0.4702 -1.5504 -0.8535
#> -0.2414 -0.8318 -0.2588  0.7659  0.2098  0.8641 -0.0533 -0.2513  0.8407  0.8943
#>  0.0712 -1.6151 -1.9372 -0.1961  2.3861  1.8245 -1.7517  1.0010  1.4452  0.6638
#>  0.3101 -0.4957 -0.2098 -0.1192  2.4636 -1.5653 -0.5226  0.4774 -0.2751  0.0100
#> -0.3334  0.3729 -0.2052  0.3591 -0.9928  0.2614 -1.3047  0.2301  0.0662 -1.3740
#> -0.2412 -1.1163 -0.6198  2.4457  1.3428  0.1946  0.2235  0.3439  0.1045 -0.6788
#>  0.0246 -1.0708 -0.6200 -1.9987 -0.5642 -0.0687  1.2393  1.9197 -1.4744  1.2890
#> -1.7050 -0.2295  0.0464  1.1772  1.4569 -0.2381 -0.5027 -0.8322 -1.2864 -1.6056
#>  0.1851 -0.7774  1.7989  1.8629  1.2152 -0.3085  1.9266 -0.9055  1.1807 -0.0063
#>  1.1766  1.7584 -0.1802 -1.0606  1.5695  0.1063  0.3180  0.1843  1.1837 -0.5981
#> -0.8356  1.7427  0.4921  1.0019 -0.7938  0.0395 -0.3287 -0.9192  0.5418 -1.4479
#> -2.0073 -0.6592  0.9123 -0.6391 -0.3098 -0.2746  2.7885  0.1693 -0.4728 -0.9906
#> -0.7292  0.2571  0.0967 -1.0064 -0.6601  0.7398 -2.3419 -1.0630  1.1594  2.5359
#> -0.4924 -0.2214 -0.7591  1.7154 -1.4819  0.1306  0.9747 -1.9878  1.8191 -0.0814
#>  0.9112 -0.2117 -1.3526 -1.1684  0.7736  1.0244  0.1319 -1.1515  0.4167 -0.6896
#> -0.7841  1.0876 -0.1448  1.4642 -0.9716 -0.3784  0.0964  1.3590 -0.5253  1.1480
#> -1.2197  0.9974  0.8561  1.2915  0.1922 -1.2223  0.7394 -1.4434 -0.8145  0.3112
#> -1.4975 -0.7954  1.1278 -0.7604  0.4713  0.3875 -1.5471 -0.5264  1.7377 -0.4463
#> -0.0708  0.5895  0.9724 -0.4705 -1.7023  0.7761 -0.1141 -0.1169 -1.4677  1.5301
#>  1.7106 -0.0415  0.8842 -1.4566  0.6132 -1.9638 -0.3070  0.1596  1.7980  0.9543
#>  0.1217 -0.1552 -1.9682  1.3804  0.3402 -0.3984 -0.2668 -0.7056  0.4253 -0.4306
#>  0.2708  0.9358 -0.6361 -0.2920  0.2560  0.3224  0.7664 -1.2558 -1.0953  1.5375
#>  1.1460 -0.5003  0.7750 -0.4905  0.5223  0.0786  0.5123 -0.2352 -0.2089  0.2962
#>  0.9857 -1.6494 -0.5444  1.1371  0.0886 -2.2640  0.5548 -0.9853  0.2726 -1.0219
#> -0.5928 -1.7785 -0.2248 -0.4299 -0.1806  1.0041  0.1553  0.9030 -1.2949 -0.8370
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]