Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6391  0.5402 -0.8285  0.5698 -1.3107  0.0844  0.7482  1.4250  0.7466  1.4245
#> -0.1640 -0.0334  0.0016 -0.9134  0.9059 -0.2464  0.5239  0.3526  0.8824 -1.3250
#> -0.6479 -0.2411 -0.6319  0.7465 -0.8695 -0.2291  0.0065 -1.5361  0.6955 -0.4596
#> -0.3053  0.6253 -0.2875  1.3894  0.6656 -0.3991 -0.4573 -0.3312 -1.4064  0.7706
#>  0.6204 -1.3090  1.0837 -0.6415 -1.1197 -0.5519 -0.2513 -0.6474  1.1419  0.0672
#>  0.2226 -0.9165  0.7505  0.5905  0.7790  0.6449  0.3222 -0.0402 -0.6252  0.4109
#>  0.6623  1.5715  1.3321  0.3578 -0.3627 -1.3732 -1.2907  0.5774  0.3697  0.6850
#> -0.7108  1.4822 -0.8761  0.9704 -0.9259 -0.4296 -0.7069 -0.4923 -1.6741 -1.3294
#>  0.3106  0.1690 -1.1561 -1.1921 -0.9039  1.1668 -0.7602 -0.3635  1.0394  0.8351
#> -0.2887 -1.7857 -0.7067  0.0123  0.2034  0.7186  0.8200  0.8858 -0.4906 -1.5836
#>  0.2443 -0.8515 -1.6155  0.2771 -1.3695  1.9948 -2.4629 -0.2947  0.1926  1.0353
#> -2.2792 -0.0472  0.4495 -0.1966  1.3605 -0.5839 -1.2605  1.1075 -1.4360 -1.7089
#> -0.5272  1.1043 -0.4623  2.1489  1.7504  1.1894 -0.0823 -1.1226  0.1075 -0.0598
#>  2.2885  0.2300  0.2781 -0.6093  0.1258 -0.3608 -0.1148 -0.0856  1.7983 -0.7464
#>  0.1865  1.0880 -0.3630 -0.3549 -0.8846 -0.2806 -0.8754 -1.2176 -0.4516 -0.8189
#> -1.1821  1.0412  0.4121  0.4757  0.6911  0.2626  2.3272  1.3985 -0.3800 -0.7178
#> -1.6317 -0.7525 -0.1230  0.3325  0.8761 -2.6447 -1.2962 -0.3366 -0.6253 -0.4224
#> -0.5911  0.7178 -0.7548  0.2250  1.0488 -0.3343 -0.5381  1.3602 -0.4860 -0.8757
#>  0.2222 -0.9047 -0.1948 -0.2842 -0.6519  0.4430  0.3184 -1.6791 -0.1032  0.2915
#> -0.4170  0.1455  0.0861 -0.3797  0.9459  0.0255  0.5729  0.5703  0.0163 -1.8550
#> -0.5434 -2.2868 -0.6045  1.5133 -0.7967 -0.5086 -0.1153  0.5520  0.4569  0.4487
#>  1.3328  0.9657 -0.3398 -1.6076  0.1585  0.2350  1.3723  0.5056 -0.6748 -0.4594
#>  0.2619 -0.2807  2.1882 -0.7082 -0.8159  0.7059  1.3826 -0.0218 -0.1951  0.5060
#> -0.1402 -0.2841  2.1526  0.5092  0.9492 -0.0536 -1.1784 -1.1090 -0.4784 -0.5768
#> -0.1459 -0.4797  3.0747 -1.0525  0.2171 -0.4258  1.1497 -0.4697  1.2287  1.9636
#> -1.4671  1.5000 -0.9019  0.7117  0.3101 -0.1533  1.0008 -0.9695  0.6922 -1.4891
#>  0.4537  0.6453 -0.0279 -0.4009  0.2557  1.0674 -1.8000 -0.5394  0.6280  0.1752
#> -0.1441  1.5644  0.3308 -2.0033  0.9364  1.0686  0.1336  0.6069 -0.5594 -0.5650
#>  0.2769  0.6660 -1.6783 -0.0736 -0.8121 -0.7936  0.1150  2.2605 -1.2532 -0.7324
#> -1.5754 -1.0032 -0.3629  0.9196  0.5752 -0.1574 -0.5138  1.2329  0.6573 -1.2756
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]