Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2755 -2.0749 -1.9700 -1.9412  0.5879  0.5363 -0.1452  0.4035  0.1799 -0.1847
#> -0.0517 -1.0333 -0.6151 -0.9236 -0.2380  0.3057  1.4191  2.0571 -0.2214  0.2369
#> -1.0208  0.7177 -0.2792  0.4103  0.4094  0.4212 -1.0162 -0.3525  0.7248  0.6858
#>  0.1540 -0.7418 -0.8251 -0.5169 -0.5296  0.1404  0.4198 -0.7869  0.3670 -2.3159
#> -0.9370  0.7961 -0.1636  1.7236  0.7573  1.8811  0.4252  0.7972  0.3110  1.1239
#>  0.4942  0.1022  2.5530  0.0392 -1.0525  1.4180 -1.7124 -0.6589 -0.7260 -0.8533
#> -0.3990 -1.0384  1.0301  0.7482  0.4998 -2.5151  0.9938 -0.4804 -0.9288  0.2899
#>  0.0047  1.3026  1.5124 -0.1224  1.3352 -0.6644  0.4678 -1.5520 -1.3866 -0.8927
#>  0.0203 -1.1498  0.9231 -1.6363  1.3500 -0.0311  1.3560  0.5497  0.8358 -1.2825
#>  0.2523  1.4148 -0.3830 -0.3291  0.1702 -0.1266  0.5749 -0.0498 -2.0029 -1.1410
#>  2.2593 -0.2913 -0.2847  0.2775  2.9528 -0.1849 -0.4511  0.2532  2.0955 -0.1150
#>  0.1214  0.1299 -1.1379 -0.4309 -1.3548  1.0279  0.5266  0.0387  1.4979  0.6307
#> -0.4251 -0.6060  1.1181 -0.0850 -0.2037  0.4498  2.1072  0.9295 -0.8304 -0.4975
#> -1.4754  0.8676 -2.1514 -2.0395  1.2787 -0.2513  0.8026 -3.3823  0.9963 -0.1399
#>  0.1659  0.9245 -0.4319 -1.2832  0.1364 -0.7198  0.5337 -0.7395  0.1444 -2.2081
#> -0.4035 -0.3690  0.0020  0.2923  0.5128  1.4394  0.7773  2.8640  1.9840 -0.3267
#> -1.0554 -0.0137  0.2404 -1.3319  0.8697  0.2362  1.0234 -1.1704  0.5413  0.1818
#> -0.7715  0.5891  0.8420  0.3972  0.7787 -2.2105 -0.9534  0.1437  0.2334 -0.2105
#> -0.8926 -1.8256  0.0872  0.2585 -2.9646 -0.7094 -0.5457 -0.1947  1.0979 -0.1341
#> -1.5561 -0.7953 -1.1149 -1.2833  0.8175 -0.3512  0.2210 -0.9666  1.6723 -0.0880
#> -0.5203 -2.2545 -0.2671 -1.3664 -0.8319  1.0387  0.9491 -0.8105  1.1095 -0.1343
#> -0.4404 -0.8308 -1.6667  1.2269  0.2413  0.2837  0.7640 -1.5368  0.1300 -0.1664
#>  2.1188  0.4269 -0.6767 -0.0275  1.9924  0.5640 -1.0100  0.2544  0.4065 -0.6155
#>  0.5677  0.0299  0.3047 -0.1234 -0.7835 -0.7245  2.0271 -1.5685 -1.0423  0.6837
#> -0.3006  0.6884  0.8444  0.9008  0.7148  0.0171 -0.9691  0.0849  1.2271 -0.2143
#> -2.4798  0.6464  2.1833 -0.0226  1.4864 -1.0079  0.5776 -0.1111 -1.2516  1.5364
#> -1.5458 -0.7709  0.7506 -1.1647 -0.2657  0.7431  0.8528  0.9217 -1.9533 -0.4816
#>  1.4451  0.1506 -0.9057 -0.2853 -1.6022 -0.2837  0.6932 -1.3034 -0.4802  0.2469
#> -0.3175  0.2626 -0.0493 -1.3074 -1.6992  0.6687 -0.7043 -0.2062  0.2470  0.7700
#> -0.1785  0.8489 -1.4247  0.9263  0.4428 -1.4298  0.8833 -2.8638  0.7889 -0.4450
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]