Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.7346  1.3318  0.6671  0.7319 -0.5470 -0.6462  1.6522 -1.5296 -0.2351  0.4842
#> -1.7600  0.8038 -0.8287 -3.0731 -1.0054  0.3780 -0.3316 -0.6104 -0.6415  0.8984
#> -0.7847 -0.1549  0.1693 -1.2916 -0.7983  1.5616  0.7495  0.0770  0.9775  2.1812
#> -0.6622  0.2406  0.9548  0.2090 -0.3496 -0.2673  0.2366  2.3971 -0.9227  0.5379
#> -1.9064  0.3799  0.7178  0.2800 -0.3819 -0.2811 -0.6788 -0.2821 -1.0908  0.5228
#>  0.2554 -0.3292  1.7665  0.4687 -1.5072 -0.2285 -1.7102  0.1764 -0.1523  0.1417
#>  0.2508 -1.0558  0.9236  0.6447  2.2197 -0.3328 -1.5915 -2.0733 -0.6731 -0.1594
#>  0.0581 -0.8718  0.9243 -0.5392 -1.0181 -1.2481 -0.8213  0.2164  0.1347 -1.7871
#>  1.7656 -0.4499 -0.0141  0.1477  0.2066 -1.5704 -0.8273 -1.0104  0.0868 -0.4387
#> -0.9515  1.5777 -0.0033 -0.4603 -0.5031 -0.5588 -0.6189  0.4340 -0.3971  0.9125
#> -0.3330 -0.8989  1.0552 -2.1092 -2.4294 -1.2564 -0.5205 -1.1293  0.4256  0.9306
#>  0.9718 -0.9802 -0.5005 -0.4609  0.3876 -0.6011 -0.8863  0.6917 -1.4373 -0.1855
#>  0.2941 -0.0031  1.1904 -0.2345 -0.0787  0.3361 -0.3318 -0.1798  0.1075 -0.7480
#>  2.0069  1.4908 -0.0818 -0.3389 -1.1755 -0.8658 -0.9817 -0.6308 -0.1428  0.3002
#>  0.3243  0.8368 -0.7311  0.2917 -1.0076 -1.5962  0.0670  0.1088 -0.7113 -0.4684
#>  1.2902  1.2366  1.9567  1.2400 -0.0988  0.4461 -0.1385 -0.1243  0.2905 -0.2186
#>  2.6529  0.4821 -1.4639  0.7642  1.5650 -1.3622  0.9646  0.5268  0.6596 -0.1737
#> -1.3802  0.1049  0.9518 -1.6446  0.3607 -1.7465 -0.5924  0.5533  1.2069  0.1382
#> -0.3083  0.3171 -0.0518  0.1089 -0.4534  0.0265 -1.9396  0.6740 -1.0739 -1.9070
#>  0.9169 -0.8279  1.7933  3.0923  0.3114  0.2532  1.1682  0.4240 -0.9420 -0.8861
#>  1.4870  1.0403 -0.7984  0.2131 -0.1851  1.5335 -1.3129  0.0273  0.0655 -0.0121
#> -1.8226 -1.1920  1.2371 -2.3823  0.3425 -0.7935  0.3111  2.3417 -0.8295 -0.0785
#> -0.3692 -0.9751 -0.1761  0.1929 -1.1126  1.0889 -1.0953  0.1462  0.5336 -0.7160
#> -0.1460  0.8611  0.3278  0.5566  1.1499  0.6497 -0.0117 -2.3361 -0.6373 -1.7086
#> -2.9055 -0.2452  0.5855 -1.4788 -0.9058  2.2554 -0.4693 -2.1972  0.1036  0.1579
#>  0.4837  1.6594  0.6696  2.5652 -0.0273  0.9844  0.2006 -0.0656  0.7539  0.4491
#>  0.9758 -0.1715 -1.3107  0.8149  0.7384  0.4770  0.9183  0.3049  1.0873  1.7368
#> -0.3398 -0.4254  1.6327  0.9385  0.1064 -0.5902  0.3603 -1.0328 -0.7889 -0.4443
#>  0.4646 -0.3559 -0.3208 -1.1342  1.4620 -0.4343  1.7429 -1.1746  0.0573  0.0367
#>  0.2048 -0.1415 -0.1616 -1.8800  1.8942  2.1309 -1.4141 -1.1093 -0.0678 -1.6419
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]