Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4565  0.3765 -0.9688 -0.3256  0.4919  1.9514 -0.0653 -2.8063 -0.2881  1.0078
#>  0.3929 -0.1530 -1.6562  0.3446  0.9232  0.9794  0.2210 -0.6396  0.4190 -0.3948
#> -1.1186  3.3851 -0.8119  0.4279 -2.0181 -0.0223 -0.2579  0.0687 -0.1082 -0.4513
#> -1.4112  0.6602  1.3129  1.2467 -0.6475  1.0443 -0.7250 -0.1156  0.3832 -0.2126
#>  0.9875 -0.0885 -0.6186 -0.4830 -0.1535 -0.4403 -0.8262 -0.8256 -0.4705  1.1280
#> -1.4066  0.9082  0.7358  0.6381 -0.8632  0.3246  0.9491  0.5477 -0.7021  2.5070
#> -0.5671  0.6107 -0.5551  0.2112  0.2279  0.1829  1.6178  1.3450  0.5702 -0.7474
#>  0.3892  0.6353 -0.8796  0.2249 -1.0227 -1.1380 -0.1495  0.3055 -0.6705 -0.2293
#>  1.4465  2.1051  1.1398  0.0328 -1.2279 -0.4193 -0.4032 -0.5807 -0.3874  0.4020
#>  0.3183  0.9282 -1.3676  0.3737 -0.4690 -0.3276  0.9964  2.1053 -2.2779 -1.0820
#>  0.6650 -0.1411 -1.4106  1.1890  1.8685 -0.9402 -1.8955  1.8448 -1.5083  0.0721
#> -1.6325  0.0248  0.5608  1.0174 -0.6556  0.8041 -0.2525 -1.4042  1.2607  0.9129
#> -0.5960  2.5681 -1.6010 -0.1118  0.0669 -0.3542  0.2264  0.5341  1.1750  0.1416
#> -1.1273  0.1260 -1.4462 -1.1322  0.8505 -0.0287 -0.6198  0.9758  1.0732 -2.3039
#>  1.7793  2.5436  0.2551 -0.0978 -0.8698 -0.0846 -0.8125 -0.8528  1.0610  0.0784
#> -1.0754  1.2390 -0.9694  0.7940 -1.1968 -1.3428 -0.2152  0.4537  0.4302 -0.6223
#>  0.7306  0.2688 -0.1571 -0.9864  0.8664 -0.0251  0.4854  0.3842 -1.3998  0.5309
#>  0.0564 -0.6734  0.2506  0.8315  1.0553  1.2906  0.7149 -1.6884 -1.4602  0.1341
#> -0.9845  1.1648  0.0401 -0.7446 -2.0092 -0.5399 -1.6220  1.9998  1.5274  0.7661
#> -0.7961  0.6401 -0.6366 -1.1178 -0.8234  1.3484 -0.2348 -0.2674 -2.5918  1.3064
#> -1.6234  1.1126  0.7495 -0.2732 -0.1790  0.3174  0.8927  1.4339 -1.1213 -0.4030
#>  0.5444 -0.1749 -0.3545  0.3104  0.9474  2.7194  0.5613  0.9104  0.8027 -0.2652
#>  1.1122 -1.7166  1.0998 -0.3664 -0.1691 -1.5347  1.8677  0.5520 -1.1595  0.9274
#>  0.1844 -0.1210  0.5176  2.1401 -1.0376 -2.5362  0.1784  0.6198 -0.0811  1.4767
#> -0.9790  0.3441 -0.6340  1.1632 -0.6398  0.6564 -1.8023  0.6949  0.9112 -1.2763
#> -0.5407  0.3939 -0.9689 -0.5121 -1.8762  0.7825  0.1969 -0.3181 -1.5095 -1.6130
#>  0.6403  0.2867 -0.9029  0.4605 -0.4065  0.8475 -0.8936  0.4920  1.4416  0.1264
#> -0.4262 -0.3069  0.7231 -1.0361  0.5919 -0.2888  0.0217  1.4872 -0.1993 -1.1980
#>  0.5920 -0.5878 -1.0107 -0.7493 -0.2970  0.3136  0.1567 -0.4399 -0.6928  0.6308
#> -0.6420  0.2179  1.0883 -0.2849 -0.8244 -0.1058 -0.4990  0.6664  0.4843 -0.9644
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]