Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6540  1.3071 -0.9622 -0.6183 -2.1093 -0.7736  0.1418  0.3154 -2.0384  0.8391
#>  0.0595 -2.0990  1.9084 -1.0943 -0.5144  0.6107  2.7933  0.4208  0.5777  0.2772
#>  0.7508 -0.2546  0.6164  0.6805  0.0997 -0.2730 -2.4397 -1.6027  0.3134 -0.4227
#>  0.5237  0.7735  0.1813 -0.0402  0.7439  1.4332  1.3222 -0.8912  0.6284  1.1033
#>  0.8605 -0.3241 -1.4266  1.0407 -1.2179 -0.0030  1.1463  0.1853 -0.6122  0.3124
#>  0.3071 -0.4310 -0.1516 -0.6864  1.2684 -0.7887  2.1598 -1.2154  0.8535 -0.1882
#> -0.4119  0.8256  1.0357  0.9232 -2.2155 -0.4025  0.6003  1.4015  1.6084  0.7193
#>  1.9853 -1.1697  0.3406 -0.9143 -0.1028 -0.0958  0.9266  1.3425 -0.2090 -0.6761
#>  0.1939 -0.7531  0.8768  0.4651  0.2654 -1.3180 -2.4193  0.1712  1.1328 -0.9460
#> -0.8777 -0.5649 -1.7571 -1.1189 -0.9600 -0.9329 -0.5931  1.3484 -0.9926  0.4106
#>  0.5916  0.2446  0.9507 -1.1527 -0.0664  0.8814  0.3525 -0.8087 -0.3841 -0.1267
#>  2.0219 -0.6144  0.9797  0.1462  0.6945  0.2229  2.7589 -1.6217 -0.3282 -1.6018
#>  0.0798  0.8365  0.3055 -0.1246 -0.4293  0.5573  0.2532  1.3103  1.5562  0.9411
#>  0.3741 -0.2474 -0.1706 -0.5901  0.3579  0.8283  0.1408  0.7483  0.0711 -0.4326
#> -0.4695  0.1631 -0.8206 -0.1862  0.7331  0.1730 -1.2364 -0.4088 -1.2627  0.7666
#>  0.7664 -0.6861 -1.4098 -1.0072 -1.3288  1.1909  1.4496 -0.6919  1.0160 -0.3612
#> -1.5507  0.4350 -0.4500 -0.6783  1.6855  1.1443  0.1238  0.7320  0.3387  0.3573
#> -0.4841  0.4375  1.0866 -0.1889  1.6143  0.6754  0.7398  0.9778  0.0099 -1.1685
#>  1.6916  0.1076 -0.9062 -0.7118 -0.2491 -0.3646  1.1858  1.1608 -1.0344 -0.5179
#>  0.6309  1.6380  0.2709  1.1380  0.3602 -0.7051 -2.1384  0.3190  1.3508 -0.4027
#>  1.5230 -1.7493  0.3926 -0.6289  0.7072  1.1586 -1.0054  1.3575  0.4285 -0.4610
#>  0.4354 -0.9325 -0.0229 -0.9967  1.5006 -0.8488 -0.3232 -0.5525 -0.9961 -1.0216
#>  0.2808  0.6311 -0.8798 -0.6184  0.2637 -1.5781  0.6163 -1.2244 -1.7185 -0.1116
#> -0.1836 -1.2155  0.4499 -0.1033 -0.9725  2.4032  1.0026 -1.6394  0.5492 -3.0115
#>  0.4251  0.0169  0.3859  0.6443 -0.4273  1.5555 -0.0236 -1.3660  0.4006  1.1649
#>  0.2155  1.5584 -0.8332  0.9300  0.1978 -0.9230 -0.8662  0.3909  1.4929  0.0094
#>  0.7850 -0.7738  1.8992  1.1442  1.9969  0.7922 -0.5593  0.5909  0.0432 -0.9011
#>  0.8716 -0.3839 -0.7163 -1.0899 -0.2094  1.1132 -0.1563 -0.8293 -3.3857  0.3708
#>  0.6168 -0.4264 -0.3026  0.6867 -1.1131 -1.4427 -2.0492 -0.8641  1.6051 -0.0690
#>  0.2808 -0.4106  0.2228  0.4826 -0.2027 -1.1695 -0.6167  2.1720 -0.4834 -1.3161
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]