Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5148  1.1433 -0.7802 -0.0459 -1.0493 -0.1862  1.0504  0.7972  0.5158  0.1306
#> -0.2650  0.8079  1.0015  1.8496 -0.3586  0.4504 -0.8076  0.1656 -0.7839 -0.5685
#>  1.9964  2.9630 -0.4788 -0.6571  0.6587  1.2629  0.1072 -0.5517  0.1732  0.5285
#> -0.1657  0.6536  0.3859 -1.4472 -0.3232 -1.1243 -1.5588 -0.8091  1.0279  0.9205
#>  1.1830 -0.0395 -0.9964  0.3518  0.2794 -0.9696  0.0412  0.9813 -1.3130 -0.6076
#>  0.4882 -3.1081 -0.1762  0.6983 -2.0736 -0.2914 -0.1539  0.2053 -0.8612 -1.3515
#>  1.2347  1.4834  0.1343 -0.4582  0.1441  1.2919 -0.2235  1.8744  0.0400  0.2760
#>  0.4013  0.6380 -0.0278 -1.2895 -0.2371  0.6096 -0.5252 -0.1218  0.4764  2.0029
#>  0.7163  1.6972 -1.3130 -0.8232 -1.4428 -0.6255 -0.1173  0.3057 -0.6906 -1.4969
#> -0.4142  1.2469  0.2324  0.3293  0.6554 -0.5838 -1.5952  0.8363 -0.5159 -1.9570
#> -0.1833  0.1813 -0.3141  0.0153  1.5587 -0.9425  0.6265  1.2610 -1.1253 -1.2307
#>  1.6937 -1.3779 -0.3580 -1.7555  0.3006  0.3015 -0.3813  1.6001  1.0210  0.8038
#>  1.0833 -1.0669  0.0683  0.7302  0.8712 -0.7837  1.6651  1.5111 -0.6689 -1.6281
#>  0.7479  0.5595 -0.2310 -1.2536  0.0597 -0.3540 -0.6249  0.0761  0.7405 -0.2949
#>  0.2208  0.0403 -0.2638 -0.9451  0.4661 -0.2547 -0.9639 -1.2741 -0.4534 -1.3653
#>  1.1363  0.4341 -0.2857 -0.8892 -0.2300  0.2622 -1.4560  0.3413 -0.7919 -0.0237
#>  1.1592 -0.3009  0.0983  0.5641 -0.1276 -1.1315 -0.0745  0.7118 -1.4776  1.6976
#> -1.4534  0.7231 -0.9521  0.3034  1.5579  0.5371 -0.7946  0.1292 -0.1725  0.6559
#>  0.2593  0.7947 -0.8253 -0.1320  0.2789  1.7766 -0.0992  1.1970 -1.5086  0.8733
#> -1.0418 -0.2488  0.3939 -2.5343  0.1021  1.8923  2.0880 -0.0707 -0.5197  0.9248
#>  0.4373  0.6186 -0.2654  1.3001  1.3918  0.2304  0.2111 -2.6398 -1.4074 -0.2617
#>  0.1313 -0.4625  1.5305  0.4489  1.2383 -1.2204 -0.7439 -1.2112 -0.3194 -0.4174
#> -0.7509  0.6354  0.1257 -1.0874 -1.2379  0.0320  0.2084 -0.5259  0.5242 -1.6256
#> -0.1217 -0.7717  0.1974 -0.8022  2.2679  1.6411  1.0669  0.3819  0.1135  0.7481
#>  0.9436  1.8965 -0.6799 -0.2527 -0.4253 -2.4765 -0.0127 -0.3055  1.9060 -0.7396
#> -1.3930  0.2024  0.4678 -0.6055 -0.9998  1.4682 -1.3874 -0.4258 -0.3422 -0.8941
#> -0.8010 -0.1113 -0.7761 -2.3982 -0.8154  0.0271  1.5754 -0.2302  0.2515  0.1165
#>  0.0934  0.5294 -0.4655  0.5204 -0.6867  0.5629  0.2725 -1.8651  1.4436 -1.3868
#> -1.1298 -1.2670 -0.9654 -0.7232  0.2041  0.4571 -1.6154 -0.1418 -1.6383  0.1024
#>  0.2858 -1.6072  1.8406 -1.3978  1.6349 -0.7460  0.0066  0.2561  0.6843 -0.4380
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]