Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.5466  0.1322 -1.3272  1.0390  0.7884  0.0399  1.6027  2.0677 -1.5352  1.8201
#>  1.5042  2.0400  0.0585  0.6740  1.0866  0.4925  0.8450 -0.4180 -0.3426 -0.6527
#> -0.7866 -0.9629 -0.9577 -1.6634  2.0566  0.2924 -1.5092  0.2023  1.0730 -1.0886
#> -1.5296 -0.7048  0.6160  0.6005 -0.5435 -1.5007  0.2700 -2.7283  1.0952  0.4961
#>  0.5575  0.7775 -0.4213  0.0652  0.5500  1.0807 -2.5973  0.7387 -0.2027  0.7668
#>  0.6784  0.7686  0.6675 -0.3041  0.7876  3.3184  1.5208 -0.4357 -0.6000 -0.2232
#>  1.8010 -0.7048  0.8518  2.5157  0.6113  0.5841 -0.0917 -0.7706 -1.2650 -0.3592
#>  0.7687  0.8379  0.7193 -1.8107 -2.2618  2.1499 -0.4764  1.0939  0.7323  1.0577
#> -2.3198  0.4741 -0.2863 -0.9481  0.3001 -0.0809  0.0424  0.9616 -1.0661  0.5418
#>  0.6152 -0.2353  0.6570  0.8128 -0.5074 -0.5995 -1.6282  0.0208  0.0647  0.6035
#>  0.4583 -1.6287 -2.5385  1.2531 -0.8198 -1.4760 -0.2348 -1.0508 -0.9721  1.2997
#>  1.7193 -1.6392 -0.7457 -0.8964  1.3409 -0.3892 -0.5870 -1.3232  1.5853 -1.9445
#> -1.1982  0.6957 -1.8654  0.2064  0.7593  2.6394 -1.1920  0.2490 -1.8487  0.3251
#>  0.3095 -3.1179 -0.0314 -0.2358  2.5368 -0.9668 -0.1911  1.0068 -1.2782  0.5627
#> -0.8209 -1.4481  1.2236  0.3623  0.0718  0.4958 -0.7051  0.6451  0.5063 -0.5634
#>  0.5187 -0.5940 -0.1976 -0.8572  1.9503  0.0717  0.1039  1.0295  0.8497 -1.6588
#> -0.8872  1.1025  0.6434  1.3472 -1.6146 -0.0254 -0.5375 -0.8096  0.6560  0.5851
#> -0.0804 -0.4602  0.2235 -1.6578  0.6081  0.1508 -1.4443 -0.5223 -0.8764 -0.6514
#>  0.4337 -1.5764  0.9544  1.7648 -1.2334 -2.2061 -1.4587 -0.6822  0.4368  2.3367
#>  0.3074 -0.9535  0.3098 -0.8657 -0.2547  0.9888 -0.1335 -1.2832  0.3955 -0.3932
#> -1.2853  1.6844 -0.1619  1.1042  0.1488  0.4942  0.9541  0.5401 -1.1764  0.3840
#>  0.1192 -1.7448  0.9775  0.6655 -1.0148 -0.9013 -0.7580 -1.3820  0.0158 -1.4462
#> -0.2686  1.2096 -0.3182  0.4505 -1.0865 -0.5731 -0.3508 -1.4852  0.5336 -0.8706
#> -0.5522 -0.5151  0.1486  1.2907  1.6727  0.2046  0.4032 -1.1590  0.0715 -0.4647
#>  2.5168  0.1999 -2.0400  0.7335  0.2110  0.0847  0.3828 -0.7397 -0.3818 -0.3474
#>  1.6703  0.8106 -0.8512 -0.0991 -0.2998 -0.8231  1.5084 -0.1903 -0.1268 -0.7244
#>  0.4035 -0.7536  0.1310 -0.7775  0.4557 -2.0453 -0.3249 -0.5558  0.3395 -0.5639
#> -0.4940 -0.4231 -1.1214  0.0851 -0.6039  0.3315  1.6724  0.1126  0.1917 -2.1111
#> -1.6529 -1.3828 -0.3374  0.2746 -0.7951  1.5226  0.1521 -0.1176 -0.4410  0.6317
#>  0.3075  0.9034 -0.0503 -0.5354 -0.3355  0.7019 -0.0318  0.4491  0.9031  1.0349
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]