Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.7924  1.2103 -0.0068  0.7906 -0.5939  0.3935 -0.5426 -1.2507 -1.7146  0.1900
#>  1.0004 -1.0499 -1.4590 -1.0318 -0.9116 -1.7802  1.0094  1.5903 -1.1060 -1.0413
#>  1.9731 -1.0401 -0.5418 -0.6904 -0.7075  1.1888  0.7661  1.7323 -0.5912 -0.1850
#> -0.7293  1.2692 -0.8747 -0.3843 -0.4971  1.4779 -1.7221  0.3798 -0.9881 -1.6642
#>  0.4796 -2.0208 -0.1838  0.0745 -0.9199 -1.2043  0.7768 -0.6528  0.6265 -0.7505
#> -1.0254 -0.8790 -0.1085 -0.8536  0.0093 -0.6228 -0.8434 -0.8243  0.6801 -0.3070
#>  0.8666 -0.4337 -2.1698 -0.8742  0.1422 -1.9406 -0.7354  1.1340 -0.7148 -0.9932
#> -0.7040  1.2730  2.4408 -1.8615 -0.5343 -0.7260  1.0640 -0.5296  0.1349  0.5258
#>  2.6818 -0.3424 -1.1118  0.3294 -1.2069  1.0999  0.6103 -0.4874 -0.4260  0.3697
#> -0.4150  1.7913  1.1638 -0.8098  0.4369 -0.5894  1.0049 -0.2350  0.5617  0.6470
#>  0.0994  0.6412 -0.5067  0.8789 -0.3998  0.7624 -0.9234 -0.8608 -0.3047 -0.7429
#>  0.1104  1.8208 -1.2989  1.2859  0.6406  0.1274 -0.5281 -0.8802  0.6974  1.3364
#> -1.0122 -0.6400 -1.2692 -0.1291 -0.1155 -0.8563  0.4163 -0.4104  1.0401 -2.5254
#>  1.1382 -0.8327  1.2910  1.4225  0.5205 -0.7494 -0.9082  0.0888 -0.4785 -0.8907
#>  0.0481  2.9987 -1.2840 -1.1101 -1.1828 -0.4870 -1.6989  0.3955  0.4338  0.0318
#>  2.2888 -0.1148 -1.4609  1.3330 -1.2434  0.2507  0.3142 -0.8324 -0.2495  0.6919
#> -0.0168  0.1532 -0.3071 -0.4220  1.2313  0.1889 -1.5277  0.8449 -2.2389 -1.0881
#> -1.8448  1.2353 -1.1573  0.1456 -1.5109  1.0406  0.5483 -0.1628  0.0456 -1.4977
#> -1.3256 -0.6103 -0.0963 -0.4989 -0.0793  0.1777 -1.9074 -1.1816 -0.0817 -0.8614
#>  0.1382  1.7282  2.2900 -0.3803  0.2135 -1.5565  0.9923 -0.8494 -0.4222  0.5894
#>  1.7548  1.4000 -0.2241 -1.5855 -0.7954 -1.0472 -1.6019 -0.8878 -1.0932 -0.0637
#> -0.6912 -0.9375 -0.1120  1.4165 -0.0439  0.7743 -0.9464  0.4885  0.6636 -0.4238
#>  0.2280  0.5517 -0.2457 -0.1133 -0.9617  0.1800 -0.7780  0.7631  1.8160 -1.7558
#> -2.5542 -0.3514 -0.0358  0.2562  1.0011  0.3283 -0.3148  0.4970 -0.9087  0.4150
#>  2.3557 -0.3169  1.5933 -1.0899  0.6991  0.3315 -0.8409 -0.2937 -0.5020 -0.2487
#> -0.4035  0.5887  1.2836  0.1187  0.2863  0.6651  0.6202 -0.1290  1.9603 -0.5513
#>  0.2289 -1.0829  0.2114 -0.8431  0.7447  0.4780  0.7679  1.2566 -0.8301  0.4775
#>  0.8278  1.2027  0.5810  1.5831 -1.1071 -0.9625  0.1290 -1.0050 -0.1070  0.5085
#>  0.8165 -1.2476 -0.0699  1.0624 -0.2832  0.8900 -1.0673  2.2820 -0.4828  0.4935
#> -0.4200  0.7112 -1.0990  0.2890 -0.3725 -0.6989  0.9010  0.0208  1.9702 -0.5561
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]