Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2402  0.6003 -0.7024  0.4553 -0.3016 -0.3732 -0.7417  1.4777  2.6187  0.3362
#> -0.6722  1.9598 -1.2432 -0.7660 -0.4518 -1.9096 -0.3080  0.7232 -1.0393 -0.6419
#>  0.0501 -0.0149 -0.5059  1.2447  1.1762  0.1906  0.0402  1.6708  0.7585 -0.4866
#> -0.4094  1.0015  0.4437  0.6864 -0.0422  0.1560  1.0360  1.2014  0.2194  0.1624
#> -0.7006 -0.2564  0.6825 -0.2507 -1.1536 -0.1012  0.3810  0.0841  0.2520 -0.6118
#>  1.0531  0.4807 -1.0902 -0.4717  0.1144 -0.1214  0.3021 -1.1182 -1.1359 -0.7832
#> -0.7405  0.0506  0.7434 -0.4219 -0.8225 -0.9615 -0.1082  0.2406 -0.0871 -0.8812
#>  1.9486 -0.4800  0.3214  0.4413 -0.1927 -0.5736 -0.6917 -0.6411 -0.7256  0.8202
#>  1.0820 -0.5489 -0.9656 -1.0673  0.0800 -1.3982  0.0948  2.5722  0.2391  0.8890
#> -0.7241 -0.0644  0.9449  0.5564  0.4288 -0.3444  0.0744 -1.8281  2.1822 -0.0590
#>  1.8534 -0.9191 -1.7506 -0.8442  1.3412 -0.3422  1.7698  0.5477 -0.1650  0.9774
#>  0.8004 -0.3516 -0.3929 -2.4281 -0.1796 -1.6582 -0.5703 -0.1449  0.7773  0.5027
#> -1.3612 -1.9840  0.8970  0.3520  0.2667 -0.2285 -0.2172 -0.6234  0.7874  0.7024
#> -1.1261 -0.0891 -0.1668 -0.4254  1.1777 -1.2828  1.6708  0.2806 -0.8376 -0.9474
#>  0.8849  0.3198 -0.6059  1.0026  1.1100 -0.8071 -0.5938  0.5188 -1.9129  0.3983
#>  0.5206 -1.2128 -0.8449 -0.3799 -0.4796  1.0955  1.0417 -1.0307 -0.8666  0.3474
#>  0.9127 -1.8339 -0.2953  1.3415  0.2367 -0.4620 -1.0660 -0.9527 -0.0366  1.1047
#> -1.1519 -0.0335 -0.1162 -0.0233 -0.7123 -0.6426  0.1331  2.1203 -0.1502  1.4353
#> -0.4947  0.3547  1.5349  1.2161 -2.1181  0.2899  0.1964  1.2593 -0.7205  2.2458
#>  0.4507  0.3002  0.4426  1.0863  1.0537 -0.2956 -0.2189 -0.4813  0.3862 -1.3772
#>  0.2278 -0.1111  0.2944 -0.2731  0.5375 -0.2551  1.0560 -0.5819 -0.2088  0.7226
#> -1.3038  0.5529 -0.3935  0.4921 -2.3507  1.1071  0.3110 -0.7616  0.5747  0.8334
#>  0.7144 -0.5686 -0.9259 -0.6983 -0.3947  0.9854  0.2405  0.1745 -0.0522 -2.1693
#>  1.5963 -1.2547 -0.9746  0.6285 -0.2764  1.4671  0.3695 -1.1555 -0.6891 -0.4696
#> -0.9209 -0.3916 -1.0410  0.1962  1.8740  0.5136 -0.9424 -0.5128  1.5409  0.2411
#>  0.2165 -0.5233  1.0192  0.2757 -0.1838 -1.1985 -0.7775  0.0095 -1.5296 -1.2602
#>  0.0031 -0.1162  0.8225  1.4671 -0.0196 -0.4364 -1.9374 -1.5674 -1.7967 -0.7128
#>  0.0310  0.5274  1.0643  0.4260  0.9496  0.2644 -1.0897 -0.7505 -0.9114 -0.6191
#> -0.1360 -0.3680 -1.0342  2.7329 -1.2443  0.7196  0.1582  0.2979 -0.2468  1.1549
#> -1.4755 -0.4132 -0.6673  0.7907  1.4125  1.1570  0.8443  0.8932  0.0711 -0.1599
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]