Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.1709 -0.0469  0.2064  0.1269 -0.9889  0.0872  0.3877  0.9505  1.5408  1.3038
#> -1.1405 -0.7701  0.4175 -0.2200  0.5741  0.8362 -1.6447  1.4741 -0.8070  0.6866
#>  0.0177  0.1053 -0.2211  1.3416 -1.0691 -0.6231 -0.1062  0.2560  0.0119  1.4379
#>  0.3702 -1.1085  0.6595 -0.9277  0.7398 -0.3698 -1.7521 -0.6204  0.0432  0.5297
#>  0.5053 -0.2154  0.5726  0.0062 -1.2004  0.7968  1.9300 -0.1426  1.5535 -1.5957
#>  0.2330 -1.6651 -1.2454 -0.2036  2.2783 -1.3734  0.1730  1.6717  1.0007 -0.5978
#> -1.1706  1.5498 -1.2745 -0.3932  0.6271 -0.3113  0.0504 -0.8758  0.1496 -1.4186
#> -1.0229 -0.2016  1.7220  0.4306 -0.2227 -0.4134 -0.4226 -0.0567  1.0650  0.7308
#> -0.7114 -1.1342  0.2019 -1.7774 -0.0562 -1.1222  0.7532  0.1300  0.3742  0.9233
#> -0.7816  1.9486  0.2589  0.7689  0.8988 -0.4657  0.8849 -1.2579  0.3778  0.0076
#> -0.4526  0.7289  1.5592 -0.3496  1.2624  0.1698 -0.5872  2.1280  0.1921  0.0574
#>  0.8660  1.1852  0.8392  0.0737 -0.1437 -1.2955  0.5946 -0.6647  1.2791  1.7034
#> -0.2206  0.3610 -1.1653  0.0152 -1.4457  0.9524 -1.5943 -1.9291 -2.6135 -0.0786
#> -0.4247  0.4601  0.4774 -0.6398 -1.3231 -0.2920  1.5256  0.0846  0.5065 -0.0561
#> -1.3034  1.0260 -1.5194 -0.3453  1.7594 -0.2326 -1.5751  0.3380  1.4354 -0.7442
#> -1.0147 -0.6802 -0.6628  0.2079  0.8956  1.4862  1.7070  1.0124  0.9236 -1.3887
#>  0.8897 -0.1374 -0.4083 -0.6199 -0.3816 -2.0121 -1.5542 -0.0816  0.3939  0.4477
#> -0.0947  0.2503 -0.2472 -1.4308  0.7495  0.0208 -0.2433 -1.0502 -0.1943  0.6099
#> -1.4599 -0.7320 -0.3952 -1.1607 -1.1591  1.5776 -0.2516  0.5006 -0.3852  1.2795
#>  0.5478 -0.9297  0.0855 -1.2760  0.1430  1.0073 -0.7811  1.0173  1.0343 -0.1498
#>  1.2233 -0.8711  0.7970  0.2024 -2.5264  2.0392  0.1955 -0.5092  0.6756  0.2813
#> -0.6090 -0.3627 -1.0085  0.0085 -0.8140  0.0584 -0.3050  0.7877  1.4873  1.3175
#>  1.1753 -0.4022 -0.9340 -0.4073 -0.3619 -1.2614  0.2427  2.0536 -1.9313 -0.7766
#>  1.8089  0.3044  0.5252  0.7842  0.6970 -0.2143 -0.3153 -0.3221 -0.0578  0.5012
#>  1.7063 -1.3136 -0.1969  1.4907  0.7734 -1.5676 -1.2173  1.4180 -0.1798 -0.0651
#>  0.3714 -0.1262 -1.6063 -1.1495 -0.0958 -0.3529  0.3003  1.7307 -0.8651 -0.2317
#>  1.1734 -0.5711  0.3823  0.3945  0.0504 -1.6946  1.1222  0.5041  0.5847  1.3857
#>  1.3852  0.1411  0.8823 -0.3989  0.5869 -0.6326  0.8556  0.9091 -0.8516  1.8982
#> -0.1389 -0.8008  1.0286  0.3231  0.0265  0.0089 -1.9697 -0.1071 -2.1660 -2.7389
#>  0.3526 -0.3056 -0.5890 -1.4805  0.7065 -1.1955 -0.2383  1.1069 -0.3939 -0.8684
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]