Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.0851  0.5338  0.6059 -1.4473  0.2053 -1.4482  0.0850 -0.0450  1.3155 -1.2446
#>  1.3778 -1.3303 -1.6208 -0.3607 -1.3690 -0.0273  1.0876 -0.1078  0.9161  0.5349
#>  2.7796  0.0443 -0.2754  0.2373 -0.3486  0.1591 -0.3138 -0.4882  2.5928  0.8581
#> -0.6557  0.2324  1.1969  1.2871 -0.9159 -0.2134  0.7381  0.4117 -0.9725  0.6851
#> -0.1215  1.2009  1.5285 -0.2750  0.2743 -0.5837  0.1081  1.3543  0.1816  1.5895
#> -1.1671 -0.2299 -2.1415 -0.3924 -0.8834  0.6917  0.6074  0.1912  1.3620  0.4738
#>  1.0430  0.5807 -1.1437 -0.1370 -0.6067 -1.0815  1.0989  0.1112 -1.1832 -0.4306
#>  0.3824  1.8302  0.4146 -0.7580 -1.2583 -0.0321  0.1047  0.4448 -0.4006 -0.0140
#> -0.7063  0.9393 -0.8926 -1.7590 -0.8459  0.4322  0.4627 -0.5330 -1.0960  1.1339
#>  0.3760  1.1518 -0.7002  0.8276 -0.2461  2.2148 -0.0613 -1.5914  0.9143 -0.8508
#> -0.9027 -1.3790  1.4705 -0.5565 -1.8528 -0.0020  0.0004  1.4222 -0.6978 -0.8620
#> -0.3165  1.5324  2.3014  0.9907  0.4630  0.4667  0.6196 -0.9867  1.0330 -0.4438
#>  0.7107 -2.2102 -1.7343  2.8631 -0.5504 -0.6120 -1.8220  1.2554  1.4010  0.6392
#> -1.1337 -0.8901  0.5459 -1.6434  0.5893 -0.2278  0.4195 -0.1602 -1.5387  0.4210
#>  1.5119 -0.9350  0.0662  0.9096  2.2990  1.3365 -0.4130  0.8285 -2.6015  1.3395
#> -1.1677 -1.0751 -0.1716 -1.8564  1.7545 -0.2744  0.0259 -1.0135  1.0714 -0.4405
#>  0.5431 -2.3956 -0.2960  3.3025 -1.8643 -1.9908  0.0504 -1.0769  2.2613  0.4234
#> -1.3395  1.4276  0.0486 -1.8104 -1.3147  0.7365  0.2937 -1.8091  0.4435 -2.2925
#> -0.6363  0.5351 -0.1113  0.0066  0.3488  1.6012 -0.6608 -1.8543  2.4456  0.8446
#> -1.4177 -0.8447 -0.1763 -0.3829 -0.7149  1.3115 -1.7765  0.9071  1.9768  2.5313
#>  1.7764  0.2665  0.3092 -1.4390 -0.7410  1.4859 -1.4395 -0.7941  0.2408  0.2100
#>  0.7145  1.1264  1.1358  1.0948  0.8503  0.6724  0.7962 -0.7294 -0.7653 -0.5936
#>  1.4039 -0.1217 -0.5964 -0.2019  0.0355 -2.8909 -0.8108 -1.3268  0.3624 -0.9104
#>  1.4528  0.0223 -0.8899  0.3328 -0.3861 -0.9425  0.7510 -0.8117  0.4654  1.8515
#>  0.1694 -0.0893 -0.1799 -0.9442 -0.8555 -0.1832  2.2075 -2.2672  2.2500  1.0758
#> -1.1045  0.7435 -0.7110  1.4470 -1.2077  0.3208 -1.0949 -1.6650 -0.5740 -0.9519
#>  0.5565 -0.5956 -0.1968  0.5698 -0.7874  1.3159 -0.1243  0.4058  0.8976 -0.0794
#> -0.7161  0.1849 -0.0818 -0.5310  0.4825 -0.4874 -1.4538 -1.0076  0.5897  0.8925
#>  1.8686  0.1171 -1.1743 -0.7877  0.4301  0.4314 -0.2826 -1.6279 -0.4666 -0.9616
#> -0.3630 -0.9619  0.1290 -1.8608  0.6883  0.5330 -0.3169  0.0270  1.4250 -0.1502
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]