Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-2.4473  0.8568 -0.0469 -0.3601  0.5359  0.0221  0.2392  0.4118  1.2596  1.4849
#>  0.7803  2.3014  0.7808  0.9421 -0.3196  1.4157  1.2098  0.4743 -0.0628 -1.0653
#>  0.3522  0.4090 -0.3969  1.0975  0.4632 -0.1774  0.2095  0.0303  1.2236 -0.6645
#>  0.5671 -0.8509  0.1161 -0.8862 -0.1471 -0.0164  1.1629  0.2262  0.0326  0.6710
#>  0.2066 -2.3967  1.7521  0.1943  2.3602  0.8075  0.3878 -3.2087  0.4868  0.4531
#>  0.7870 -1.1072 -0.2897  1.5603  0.5558  0.0219  0.7626  0.5821 -0.4541 -1.4429
#>  0.8087 -1.1857 -0.8104 -0.4464  0.0447 -1.1703  1.4981 -0.2526 -0.3079 -0.4057
#>  1.4439  2.2143  0.1099  0.3667 -0.3062 -0.8424  1.6035 -0.0900 -0.5898  0.7500
#> -1.1338  0.7835  0.0695  0.0931  0.3447  0.8118 -1.0023  0.7657  0.6358  1.2251
#> -0.0721  1.6336 -0.6309  2.2588  0.8340 -0.2443  0.8246 -1.1286  1.9176  2.0323
#>  1.7726  0.8695  0.0163 -0.8156 -0.9587 -1.1927  0.6435 -1.2030  1.1166 -0.5851
#> -0.5550 -0.0937 -1.9226 -1.3503  0.1868  0.6575  0.7165  0.0363 -0.1253 -1.6459
#>  1.7950  0.7710  0.0696  0.8927  1.3161  0.4117 -1.4815  0.8911 -1.2198  0.5556
#>  0.2700 -0.2015  0.5558  0.1403  0.5527  0.2930 -0.7929 -0.2278  0.3352 -1.3625
#>  0.4763  0.9847 -0.6883 -0.9385 -0.4566 -0.5966 -0.0734  0.0908 -1.2762 -0.9862
#>  0.7668  0.6668  0.7010 -0.7730  0.1261  1.9483  0.9563  0.9102  0.2708  0.3866
#> -0.5718 -0.0070  0.4159 -1.3661  0.0972  0.0489  0.8848 -0.2720  1.0943  0.0909
#>  0.0465 -0.3019 -0.0128  0.9684  0.5442  0.9398 -1.0533 -0.1420 -0.9768  0.9667
#> -0.0288 -1.4759  1.6828 -1.8444  0.4878  0.8536 -0.7570 -1.0913  2.1255  0.4833
#> -1.2778 -2.9479  0.7026 -0.4356  0.3076 -0.1848 -0.0179 -0.2131 -2.5675 -1.1385
#>  0.1224  0.2576 -0.1195  1.5962 -0.6116 -0.1197  0.5575 -1.3626  0.7173  0.3969
#>  1.3581 -0.1382  0.1197  0.4836 -0.8058  0.9105  0.2159 -0.0743  1.0625  2.2075
#>  0.8778 -0.3157 -0.1448 -1.4401 -0.3973  0.4597 -1.2304 -2.4607 -1.6982  0.8550
#>  0.5708  0.3944 -0.3654 -1.4109 -1.3202  0.7607 -0.0066  0.6738  1.2634 -0.5060
#> -1.6001 -0.5965  0.6304 -0.6180  0.6493 -0.9487  0.5923  1.3974  0.9426 -0.3820
#> -0.7496  0.4219  0.7599  0.8018  1.1759  0.3994 -1.1338  0.0179  0.4140 -0.2104
#> -0.2078  1.9398  0.4897  0.0016 -0.1149  0.6221 -1.0586 -1.3929  1.4346 -1.3203
#>  0.9878 -0.2085  0.3523  0.1635 -2.3288  1.0470  0.9942  0.1070 -0.4362 -0.4550
#> -1.5406 -1.7636  1.3192 -0.1555 -0.2876 -1.6795  0.1092 -1.1065  1.2350 -1.6544
#>  0.8981 -0.1705  0.7271  0.5554 -0.1835  0.1028 -0.1733 -1.2514  0.0892 -0.1534
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]