Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6543  1.3837  0.7247  0.9187 -0.2389 -2.1069 -1.6854  0.0379 -0.7352 -1.0695
#>  0.1462 -0.3300  0.7309  0.6725 -0.2198 -0.4019  0.0058  0.5019  0.3704 -1.3837
#> -0.7996 -2.8213  0.2470 -0.2936 -0.5426  1.8014  0.2517 -0.3172 -0.9229  1.0731
#>  1.8278  0.3416 -0.4195 -0.7903  0.3579 -1.0469  0.0467 -1.5633 -0.5899 -0.4755
#>  1.3595 -0.0089  0.5481 -1.3087 -0.3734  0.1320  1.5786 -0.4206  0.3168 -1.0805
#> -1.1664  1.2656 -1.3601  0.7045  0.0643 -0.5243  1.1405  0.4085 -1.1475 -1.0539
#>  1.2588  0.1310 -0.4530  0.5123  0.1522  0.4491  0.3234  0.2102 -0.9509 -0.9185
#>  0.2204  0.9017  0.7222  1.6994 -0.0821  1.6088 -0.9834  0.2376  0.9943  0.1046
#>  1.2973  1.1180 -0.7997  1.7544  0.3668  0.1799 -0.0692  1.4147  0.6104  1.5487
#>  0.5517  0.4021 -1.0032  1.1717 -0.5317 -1.0586 -0.1016  0.0538  0.7909 -1.8069
#> -0.4413  0.5518 -0.7155  0.2829 -2.9225 -2.4460  1.2932  1.9432  0.4421  0.6171
#>  0.3326 -1.5889  0.3135  0.0503  1.9382  1.2354  1.2259 -0.7007 -1.6384 -0.3017
#>  0.7175 -1.4544  0.4872  1.0013 -0.5738 -0.5871 -0.3509  0.6424 -1.5056 -0.6286
#>  1.1995  0.1070  0.4034 -0.5286  0.4738 -1.3826  0.9726  1.2827  0.8697 -0.8679
#>  0.0290 -0.3702  0.7497 -0.5736 -0.2464  1.5096  0.0005 -0.2385 -0.8891  0.6634
#> -2.9187  0.4103 -0.5784  0.5341  1.2604  0.0211 -0.4355  0.7373  0.2376  1.2145
#> -0.7097  0.0743 -0.6717  1.4136 -1.7861 -0.4469 -0.3046  2.0237 -0.6582 -0.1079
#>  0.7510 -0.2042 -0.7828 -0.3805  2.1038  0.0794  0.9886  1.2155 -0.9273  0.2866
#> -1.5642 -0.8089 -0.3332 -0.3483  0.6567 -0.4005 -1.8359  0.9582  0.4087 -0.8904
#> -1.1039  2.7036 -0.5834 -1.2108  1.0555 -0.8411  0.0291 -0.1949  0.2763  0.1963
#>  0.6794  1.2050 -0.4688 -1.4208 -0.1199 -0.1643 -1.1422 -0.3049 -0.1886 -0.4013
#>  1.1769 -1.8823  0.5729  1.1017 -1.5454  0.5711 -0.4291  0.7376  0.4595  0.2259
#> -0.2960 -0.1745  0.8345  0.2655  0.1526 -0.2890 -0.0693 -0.8114  0.6067 -0.8072
#> -0.6296 -1.6217 -0.5402  0.3065  1.4513 -1.3340  0.2965 -0.7227 -0.1141  0.0524
#> -0.9097  1.2466  0.3342  0.3262  0.2253 -1.2879 -0.2625 -0.7361 -1.0201  0.4996
#>  0.0265  0.7987 -1.6575 -0.5750 -0.8674 -0.3988 -0.4190  0.2308  0.1107  0.6561
#>  1.7465 -0.1866  2.7176  1.0395 -0.6569  1.7253  0.3946  0.4893 -0.2775  0.6756
#> -1.1398  0.4445 -0.8834 -1.8095  0.8245  1.4735 -0.1169  1.0206  0.6922  0.3787
#> -0.2600  1.8767 -1.5445 -1.1740  1.7404  0.1080  1.5573  1.0982 -0.6526  0.0706
#>  0.1613  1.1886  0.3938  0.8730  1.0203 -0.7832  0.4332  0.4848 -1.6148 -0.3993
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]