Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2209  0.1691 -0.5726 -0.5509  1.0616 -0.7925  0.4434 -0.6391  0.2808 -0.5841
#>  1.7319  0.9373  2.0472 -1.8017  0.2707  0.9791 -0.7468 -0.0152  0.8372 -1.1484
#>  0.4212  0.0529  0.3032  1.3362  0.2739  0.2974 -0.3516  0.0231  0.1487 -0.6196
#>  0.2063  0.3836  1.0942  1.3791 -0.5622  0.0165 -2.0230  1.4147  1.4352  1.3229
#> -0.9597 -1.4545 -0.2371 -0.6259 -0.1313  1.5963 -0.6959 -0.5643  0.4450  0.4045
#> -0.7156 -0.0775 -2.2025 -1.8376 -1.5333 -1.0774  1.2978 -0.7656 -0.4034 -0.2187
#>  0.4058  0.2174 -0.9431 -1.4168  0.9836  0.8831 -2.3167  0.7930 -0.3336  1.3413
#> -0.2532 -0.2340 -1.2670 -1.4532 -2.2514 -0.5205 -1.4934 -0.1377  1.3594 -0.1226
#>  0.3290 -1.0385 -0.1415 -1.5605  0.3415  0.0712  0.5871  0.2109  0.9882  0.9037
#>  1.6941  0.9282  1.0335  1.3774  0.6778  0.0460 -0.3756 -0.6817  0.0612 -0.3067
#> -1.3005 -0.0835  0.6489  0.3169  1.2659  0.6963 -0.0728  1.3092  0.8432  0.8982
#> -0.7474 -0.9001 -2.3944  0.8109  0.6231  0.3202  1.1805  0.3133  0.0386  1.2346
#> -1.1268 -0.8926 -1.3847 -0.7163 -0.3984  0.3920  1.4351  0.2251 -0.0248 -0.0536
#>  1.7250  0.4500 -0.6968 -0.8136 -0.7419 -0.5562 -0.6693  0.0192 -0.0846  0.4485
#> -1.6686  1.0515  0.5939  1.6293  0.5248 -1.1447  0.1775  1.2235 -0.6203 -0.6575
#>  1.7553 -0.0342  0.4025  0.4072  1.9619 -1.0492 -0.9826  0.0676  0.0379  0.6590
#>  1.5954 -1.6372 -0.8029  0.6387  1.4040  0.1266 -0.3303  1.3186 -0.2297 -1.6570
#> -0.6876  1.4403  1.5010  0.8778 -0.5585 -0.0166 -0.5175  0.7217  0.9902  0.7435
#>  0.0884  0.5209  0.2165 -2.1287 -0.4861  0.4638 -0.3235  0.0906 -0.6931  1.2157
#> -0.2443 -1.4183 -2.4675 -1.1740  0.5418 -1.5856 -2.2445 -0.0152 -0.1180  1.2824
#>  1.0113  0.4911 -0.4713  0.5968  0.7749 -0.0420 -0.2380  2.0777  0.1278  0.0104
#> -0.7184  0.1734 -0.2268  2.1745 -0.5654  1.0893 -1.3463 -0.4343 -0.4182  0.6679
#>  0.2667 -0.4900 -0.3406  0.0453  0.7400 -0.8299 -0.0636 -0.5612  1.1577  1.1954
#>  0.1757 -0.6676 -2.0738 -0.3848 -0.3517  0.3503  0.1935  0.3758  1.0588  0.8861
#> -0.5967  0.0253 -0.6891 -1.4007 -0.1274 -1.4274 -1.4473 -0.7389  0.0290 -0.2650
#> -0.2429 -0.9015  0.9598 -0.0218 -0.0551  0.8181  0.7300 -1.0585  0.1275 -0.6087
#>  0.2903 -0.2939 -1.2216 -0.9122 -0.1864 -0.9298 -0.1762 -0.1761  0.8640  1.0113
#> -1.0059  0.9807 -0.8424 -0.3272 -0.5120  0.6946 -0.7526 -0.2557  0.8534  0.6800
#>  0.6908 -0.0341 -0.8616  2.6103 -0.6153 -1.4642  0.8288 -0.5922 -1.2863 -1.4452
#> -0.4511  1.0849  1.3166 -0.8300  0.3573  1.3478 -0.9025 -0.7181 -0.9161 -0.3771
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]