Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.9825 -1.4088 -1.9176 -0.5082 -0.0305 -0.6773 -0.3953  0.1226 -0.8120 -1.2142
#>  1.4196 -1.4887 -1.3152  0.1510  0.2423  0.1334  1.4450  0.5762 -0.3142 -1.9759
#> -0.3197 -0.6777  1.1991 -0.5624 -1.3572 -0.3601  0.3226  0.7269  0.6457  0.8574
#> -0.3633  1.2267  1.1830  1.0924  1.0319  0.6775 -0.4559  1.0330  0.4718  0.4874
#> -0.8475 -1.3342 -1.0484 -1.1704  1.2875  0.7044 -1.4362 -1.9036 -1.0957  0.0712
#>  0.5815  0.0224 -0.2948 -0.3762 -0.4087 -1.0134 -1.0347  1.3154  0.5574 -0.0348
#>  0.1968  0.5343 -1.7499  0.7696 -0.0257  1.1551 -0.5003 -0.8275  1.1654 -0.3766
#> -0.4933  0.1131  0.4557 -0.0884 -0.1582  0.7184 -0.7301 -0.6044 -0.2883 -0.3593
#>  1.0659 -0.5927  1.2277  0.1259 -0.9484 -0.7358  0.6621  0.5487 -0.6085 -0.1149
#> -0.4923 -0.2333 -1.3522 -0.3331  0.5584  0.0235 -2.0171  0.7672 -0.1815  0.1103
#>  0.2433 -0.7897  1.4747 -0.2884  1.3919  0.4234 -0.8318  2.0884  0.7171 -0.5220
#> -0.1305 -0.8531  0.9285 -0.2667  0.3614  0.0281 -0.8606  1.7476  0.3912 -0.1232
#>  0.1717 -0.3593 -1.8818 -0.9787  0.0828  1.1099  0.6323  0.7662  0.6155  0.9231
#> -1.1175  0.6550 -0.9054 -0.9464 -2.9557  1.0275  1.8346  0.2443  0.2240 -0.1934
#>  1.5700  0.6247  0.8006  2.4851 -0.8481 -1.8984  0.6400 -1.3575 -0.7392 -1.2216
#>  0.3720 -1.1680 -1.0319  0.0805 -1.1138  0.0697  1.7440 -0.5185  1.4764  1.6265
#>  1.0279 -0.1089  2.0886  0.9500  0.3757  0.4102 -1.0623  0.1036 -0.1487 -0.6579
#> -0.0462 -0.9943  1.0599 -1.4234 -0.2864 -2.0388  0.6391 -0.4865 -0.3988 -0.5638
#> -0.4853  0.1056 -1.0261 -1.3843 -0.3958  1.8195 -0.6949 -1.7167  0.5191 -0.1741
#> -1.4762 -0.9744 -0.5771 -0.8973 -0.8816 -0.7515  0.2814 -1.5238 -0.6732 -0.7830
#>  2.6152  0.6871 -0.2622  0.4119  0.5989 -0.0603  0.2813 -0.1882  0.4142  0.7845
#> -0.3962 -1.4932 -2.0961  1.1155  1.1892  0.2169  0.1468 -0.9478 -0.7563  0.7492
#>  0.5470  0.0080  1.6916 -0.5949 -1.4730  1.2349  0.2625 -1.2393  0.5410 -1.4966
#> -0.9210 -0.3878 -2.3693  0.1627 -1.1940 -0.4371 -0.5699 -0.4844  0.4840  0.2835
#> -0.7633  0.5943 -0.2033  2.4873 -0.9500 -1.2416  0.3827  1.2651  0.4623 -0.4565
#>  1.9980 -0.3624  1.5899  1.0694 -0.1536 -0.3340 -0.7952 -2.2680 -1.0983  0.7759
#> -1.2783 -1.1241 -0.3304 -2.0365  1.6333 -0.5323  0.1147  2.2565  0.8952 -0.5810
#> -2.2632 -1.0706  1.4311  0.3227  1.3378  1.7216  0.2649  1.7396 -0.0044  1.4266
#>  0.5871 -0.9020 -1.5787 -0.1350  0.0150  1.2551  0.1361  0.3335  0.3017  0.8162
#>  0.3252  1.7538 -0.1594 -0.2194  0.0180  0.6912 -0.4308 -0.3786 -0.0094 -0.1735
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]