Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6085  0.8613  0.3481 -0.3675  0.2313 -0.0981 -0.7127  1.4286  1.4434 -0.9860
#>  0.8227 -0.5770 -1.1363  0.4138  1.5593 -0.0916 -1.3539  1.4799 -0.6727  0.6123
#>  0.1823 -0.9221 -0.6177  0.5927  0.8891  0.9808  0.1185 -0.5055  0.9793 -0.8259
#> -0.0102  1.0491  1.4234 -0.4611 -0.1115  0.5564 -0.7280 -1.1595 -0.5771 -0.4291
#>  0.1112 -0.3201  1.6859  1.3698  1.1336 -0.2396 -0.1596 -0.5940 -0.4368  1.0001
#> -0.1442 -2.6364 -0.5294  0.2913 -0.6958 -0.3348 -0.3223  0.5909  0.3192 -0.9901
#> -2.9493  0.2547  0.9135  0.7690 -1.8457  0.5039  0.6975  2.3867  0.1682  0.6802
#>  0.2686 -1.4784  0.2554 -0.2092  1.6299 -0.9887  0.5283  1.3051 -0.2234 -0.5455
#>  0.2603 -0.7230  1.2236 -0.9316  0.4379  0.0088 -0.7710  2.2967 -0.3802 -0.5295
#> -0.6414 -0.4856 -0.4770 -0.9278 -1.6612  0.9542  0.5672  1.1632  1.2355 -0.1801
#> -1.0861  1.1116 -0.7821 -0.2914  0.3766 -0.5690  1.4476 -0.8851 -0.8738  0.6470
#>  0.8572 -1.4046  1.0798 -1.6774 -0.1234 -2.4558  0.2456  0.9356  0.4942  0.9425
#>  0.9405 -0.8426 -0.4777 -1.3382  0.6424 -0.3547  0.2945  0.4051 -0.8541 -0.2221
#>  0.6727 -0.8538 -1.0340 -2.1544 -0.1132 -0.2352  0.4424  0.4772  0.8677  1.8106
#>  0.2040  0.4223 -0.0364 -1.5493  0.3336  1.3177 -0.7731 -0.9015 -1.1679 -0.7384
#>  0.1045 -1.2286 -0.9046 -0.0082  1.7227 -1.2335 -1.4918 -1.0192 -0.3118 -0.2630
#> -1.0867  1.2142  0.6903  0.5942 -1.1540  0.4884 -0.6168  1.1631 -0.1135  1.0792
#>  0.3217  1.0244 -0.6983  0.7687  0.1505  0.1119  0.1429 -0.6677  1.2816 -0.3755
#>  0.0039 -1.0117  0.3078  0.3375 -0.5780 -0.6196 -0.2626 -0.5186  1.4634  0.4279
#> -1.2742 -1.2795  2.2716  0.3879 -1.4197 -1.3338 -0.8219  1.1142  1.5470  0.1300
#> -1.5222 -0.6578 -0.3671 -1.1230 -1.1316  0.8888  1.9739 -0.4811  0.8721  1.0096
#> -1.5244  1.7765 -0.6060  0.4102  0.0709  1.2526  1.0227  0.0290 -1.0871  0.2520
#>  0.2658  2.6088  0.7098  0.2152  0.9387 -0.0130 -1.7235 -0.1411  0.8976 -2.0885
#> -0.8252  0.8538  0.9404 -2.2641  1.2829  0.6038  0.0197 -0.2787  0.4756 -1.0086
#> -0.8692  0.1058  0.9742  0.2150 -2.9460 -0.2209 -0.7906 -0.3002  0.5400  1.6472
#>  0.0699 -0.3506  1.8688 -1.1934  2.0787 -0.5780  0.4435  0.2490  0.8407 -0.0148
#> -0.1072 -0.0560  0.9918 -0.3609  0.0464  1.2201 -0.8500  0.0648 -0.8236  0.8311
#> -0.3653  0.9366 -0.4066 -1.0956  1.0602 -0.9418  1.1456 -0.5333  1.8241 -0.0332
#> -0.9418 -1.5206 -0.5920  1.0203 -1.2075  0.7864  0.5597  1.3069  1.2111 -1.4030
#> -1.2018  0.6136  0.2202  0.2814 -0.4520  1.6396  0.3059 -0.4581 -2.0659  0.5485
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]