Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5691 -1.6913 -0.1835  0.0372  0.2617  0.4661 -0.1890 -0.4935 -0.5050  0.9560
#> -1.1014  0.4711  1.0923 -1.0572  0.9525  1.0138 -0.9411  0.8067 -1.8847 -0.8415
#> -0.1469 -0.7950  1.3416  1.0468 -0.3726  0.6839  0.7136 -0.3485  0.8259  0.2785
#>  0.5848  1.5577  0.0363  0.8021  1.3071  0.3844 -0.9386 -1.7681  1.2481  0.8796
#> -0.2839  1.0386  1.4021 -3.0991  1.6430  0.9545  0.7394  0.2154  1.4902  0.2422
#> -0.9875  1.4022 -2.1495  0.1758  2.1677  0.6287  0.5094  0.9933  0.5065  0.5960
#> -0.5041  1.8126  1.6120  0.6231 -0.5471 -0.6303  0.8168  0.4383  1.4116  0.8106
#> -0.0892 -0.1837  1.2581 -0.9683  0.3322 -1.2304  0.7977  0.4690 -0.4046 -1.1218
#> -0.6749 -0.7313 -1.4025 -1.3749  0.5218 -0.0604  1.3400  0.1477  0.3114  0.3493
#> -0.3151 -0.3592  1.1318 -0.1800  0.6498  0.1108  0.0453 -0.8975  0.1022 -1.0307
#>  0.3920  1.3982  1.4126 -1.4701  1.8850  1.2039  1.0193 -0.3914  0.2637 -0.4480
#>  1.2256  0.0621 -1.8328 -0.7767 -0.8349 -0.5136 -0.4578 -0.7240  0.0531 -2.7912
#>  1.0317  1.5362 -0.2933 -1.4851 -0.9812  1.0017  0.8018 -2.9172  2.5368 -0.1393
#>  0.6904 -1.5357  0.0457  1.7100 -1.3979  1.1177 -1.2123 -0.8663 -0.0031  1.3943
#>  1.0672  0.8907 -2.0878  0.5884  0.3712 -0.7804 -1.6771 -1.3281 -0.2999  0.6743
#> -0.3049  0.4280 -0.7647  0.2487  0.9544 -0.2160  0.4072  0.9589  0.1238 -0.6515
#>  0.5046  0.9989  0.5220 -1.5136 -0.0020  0.0423 -0.3217 -0.3365  1.7584  1.2000
#> -0.4334  0.9017  0.9624  0.7913  1.1975  1.8271 -0.3336 -0.5987  0.5643 -1.0771
#>  0.6072  2.2169 -0.8955 -0.6531 -2.0597 -0.0602 -1.6922 -0.7944 -0.2460 -3.6719
#> -0.5454 -1.1940 -1.9202  1.4916 -0.2873 -1.2562 -0.7603 -0.5862 -0.2973  0.3829
#> -1.3092  0.7794  1.3471 -0.2268  2.1415  0.9333  1.2602  0.1283 -1.1691 -0.2410
#>  0.7432 -1.2372  0.2017  0.0968 -0.3210 -0.8182  0.9014 -1.1663 -1.0612 -0.3774
#>  0.5273  0.0127  1.2313 -0.0428 -0.8518 -1.4844 -1.7131 -0.2893  1.4263 -0.3052
#> -0.5824 -0.6555  0.5443 -0.1513 -0.2254  1.7598 -0.0958 -0.4443  0.7949  1.5663
#>  0.0201 -0.1843  0.3125  0.5354  0.6664  0.0778  0.1287 -0.6479  0.0250 -0.6179
#>  1.1915 -0.7798  0.4463  0.3677  0.3060  1.2632 -1.3599 -0.8426 -0.2144 -0.5612
#>  0.2537 -0.0037 -0.4177  0.0308 -0.2551 -0.9561 -0.0305  2.2873  0.6710 -0.6787
#> -3.1216 -1.0720 -0.5110 -0.1626  0.4567  0.4234  1.1905 -0.5059 -2.0275 -0.4628
#> -0.2270 -2.0977  0.0282 -0.1751  0.9459 -1.6538  1.1905 -2.2424 -0.2338  1.5337
#> -0.4685 -0.3592  0.1450 -0.6181  0.3535 -0.6501  0.1759 -0.2209 -0.1023  0.0300
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]