Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.1499  0.3465 -1.9290  0.5808  0.4440 -0.6077 -0.2546 -0.3799  1.8642 -0.2566
#>  0.5899 -0.8005  0.6697  0.1810  0.0862 -0.2534 -0.8288  0.2029 -0.0150  0.3712
#>  0.4275  0.1534  0.7584 -0.7804  0.1827  1.1633 -0.1038  0.4198  0.6273  0.6197
#> -1.4382 -0.0313  1.4675 -1.4636 -0.3703 -0.3613  1.1442  0.8026 -0.0716 -0.8675
#> -0.9084  0.1154  0.8471  1.5466  1.2421 -0.6149  0.6832 -1.7419  0.0889  0.0490
#> -0.8330  0.2761  0.2924  1.8872 -0.4226 -0.4493 -0.7056 -0.6030 -0.6865 -1.7214
#> -0.4597 -0.9536  0.8782  1.7646 -0.3103  0.7473  0.1248  0.3119  0.1652  0.2260
#> -1.2987  0.3689  0.1145  0.8159  0.8636 -1.5191 -0.4274 -1.4050 -0.7720  0.2712
#> -2.1183  0.2435  1.0589  0.2035  1.4343 -1.7315  0.1074 -1.8458  1.5869 -0.7750
#>  0.2077  1.3614 -1.2644  3.0786  0.1226 -2.0428 -1.1284  0.4618  0.3424 -0.1900
#> -0.0530 -0.2745  0.1239  1.0897 -0.2783  0.1694  0.4108 -0.1189 -0.4657  0.3484
#> -0.2898 -0.3146  0.0438  0.9480  0.1784 -1.2131  0.9727 -0.2330 -0.4135  1.1835
#>  1.2245 -0.6592 -1.9126  1.3633  0.4098 -1.0058 -0.3799  0.4079  0.0287 -0.2813
#>  2.4071  1.2275  0.4873  1.6666 -0.6732  1.1506  0.4124 -0.3041 -0.1270  0.4538
#> -0.5858 -0.8666 -0.6399  0.5773 -0.2976  0.7064  0.0206  0.0314  0.8955 -1.3202
#>  0.4574 -1.4024 -1.0726 -0.4312  1.4859  1.3437  1.6383  1.0530  0.6805  0.3748
#>  0.3256 -0.7128  0.9526  2.1318 -2.6590  1.6056  1.3873 -0.1849  0.1444  0.9145
#>  0.2954 -1.2996  0.1971  0.0889  0.9940  0.5405 -0.7031 -0.4762  0.5452  0.7394
#> -0.6703  0.4466 -0.3244  0.7473  0.4769  0.0386 -0.6600 -0.1727  1.9042 -0.1796
#> -0.0916  0.2120 -1.7233 -0.3264  1.6372  0.4867 -0.8021 -1.0398 -0.2162  1.4857
#> -1.7833  1.0662 -0.5782 -1.1927  0.5122 -1.0466 -1.5143 -0.5583  0.5375 -0.0156
#> -1.0065 -1.1520  0.1049  1.3021  1.2719 -0.0222 -0.6652 -0.3599  0.2152  0.4999
#>  0.4144  1.0780  0.6021  1.4274  0.4400 -0.3824 -0.2051 -0.2995  1.5388 -0.4736
#>  1.1514 -0.9377 -1.1861  0.3073 -1.2193  0.1105  0.4605  1.2285 -0.7143 -1.5752
#>  0.0285  0.4578 -0.0049  2.1760  1.8824 -0.2344  0.5508  0.8944 -0.2662 -0.4635
#> -0.0302  1.0538  1.1689  0.6006 -0.5831 -1.8352  0.9198 -0.7039  0.9056  0.1908
#>  1.3285 -0.4488  1.0899 -0.8732 -0.4066 -0.5628 -2.6742 -0.4098  0.5413 -0.5965
#> -0.4744 -0.9333  0.4701  0.4267 -0.4429  0.9967  0.2559 -0.1818  0.4554 -0.6596
#> -0.5277  0.4831  0.8408 -0.4317 -0.6266 -0.0627 -1.3584 -1.6058 -0.8903  2.3748
#> -1.5443  0.6596  0.9849 -0.0303  0.7031  0.1665  0.0429 -1.4221 -0.0600 -1.0310
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]