Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.2809  0.8890  1.0454 -0.7054  0.1721 -1.7573  1.2126  0.8811  0.3494  0.1618
#> -0.9481  0.3185  0.4238  0.1438 -1.1598 -0.6893  0.9716  0.1722 -1.5979 -0.4823
#> -0.3112  0.5555 -1.2043  0.7711 -1.6541  3.0953  0.2437  1.5236  0.3793 -0.0520
#> -0.3820  0.8578 -1.5624 -1.2347 -0.0515 -0.2821  0.1573  0.6724 -0.8434  0.0956
#> -1.3339 -0.7792 -0.0847  0.7909 -0.5185 -0.7265 -0.1595 -0.9504 -1.9619 -2.6882
#> -0.3111 -1.0046  0.7890  0.8446 -0.0642 -0.3948 -0.9941 -0.7628  1.9575 -0.3675
#> -0.2890 -1.3988 -0.4576  0.1749 -0.6632  0.8229 -0.1865 -1.9922 -1.1843  0.2039
#> -0.2566  0.9825  0.4706 -0.0374 -1.8796 -0.8373  0.6581  1.2709 -0.4734 -1.8469
#>  0.1720  1.0027 -0.9106  1.8512 -0.5667 -0.3921 -0.2893  0.8573 -1.0851 -0.7803
#>  0.5147  0.0037  0.0958  0.5099 -0.2914 -0.6840  0.6434 -0.7103 -1.0727  0.9479
#>  0.8609  0.9473 -0.5358 -0.3379  0.9507 -0.8500  0.2564 -0.0665  1.0235  2.0132
#>  0.3600 -1.7757 -1.7124 -0.5758 -1.8334  1.0567 -0.1679  1.9685 -0.2359  1.1033
#>  0.3919 -1.5534  0.0187  1.1445  0.2907  1.4143  0.3884 -0.0476  0.3975 -1.2463
#>  0.8665 -0.5809  0.8837 -0.5340 -1.0922 -1.7975 -0.0272 -0.4453  0.5328 -0.1888
#>  1.0727  0.7304 -0.3238  0.7777 -0.3090  0.2159 -2.1473 -0.4897  0.3252 -2.1135
#> -0.4584  0.5752 -1.4343 -0.1852  1.3607 -0.8553 -0.1751  0.2866 -1.5343  0.0248
#>  2.9134  1.4791  0.2040  0.9376 -0.2713  0.8238 -1.2696  2.0809 -0.7801 -1.0332
#>  0.7411  0.0646 -1.7922  0.7757 -0.6538  0.2157  0.0375 -1.1726  1.0468 -0.4519
#>  0.1498 -1.9171  0.8909 -2.1176 -0.1494 -0.0538  1.0573  1.1718  0.9874  1.7486
#>  1.2063  1.4903 -1.7937 -0.2311 -0.7942  2.2964 -0.2895 -0.1865  1.4154  0.1383
#>  1.4096  0.3741  0.9544  0.0020  0.0591 -0.1112 -2.1454 -0.1922 -0.2473  0.2533
#>  0.2150  1.1239  0.7633 -0.3546  0.8263 -1.4944  1.2562  0.5070 -0.4915 -1.9410
#> -0.4711 -0.5222  0.9326 -1.2028  0.7214  1.0884  0.6516 -2.4019 -0.2325  1.1043
#> -1.8545 -0.1830 -0.7220 -0.5108 -0.6789 -2.0409  0.6777  0.0493 -0.2084  0.1497
#> -0.4242  1.0496 -0.0612 -0.0495  0.1094  0.7509 -0.8960 -0.5494  0.6118  0.1515
#>  0.7479 -1.2862 -1.0428 -1.7473 -1.1196  0.3585 -0.8824 -0.0230 -0.1902 -1.7806
#> -1.4870 -0.5236 -0.5344  0.1816 -3.1515  1.4867 -0.0507 -0.0014  0.6665  1.8052
#>  0.3512 -0.2216  1.3745 -0.8368 -0.0944 -0.4388 -1.0301  0.0296  0.9988  0.0043
#>  0.0006  0.6908 -0.8975 -0.2434 -0.4556  0.5432  0.8430 -2.4253 -0.4894 -0.2870
#>  0.6113  2.2994 -0.0220 -0.1236 -0.5658 -2.0891  0.9381 -0.8563 -0.6717 -1.4833
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]