Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.9140  0.3882  0.3589  0.5623  0.7615  1.6008 -0.0041  0.1797  1.4585  0.9252
#> -0.8943 -1.8277  0.0302  0.7695  0.5310 -1.2614 -0.2951  2.2500 -0.6206 -0.4986
#>  0.4476  1.0339 -0.0557  0.8350 -0.9182  0.9341 -0.6651 -0.1164  2.5824  1.0158
#> -0.0177  0.3362  1.1722 -1.0001 -1.2247  1.5646  0.9761  0.1416  0.0888  1.0860
#> -2.2858  0.4004 -0.0904  1.4564  1.1559 -2.4151  1.6220  0.1496  0.4676 -1.0857
#> -0.2634 -1.8712 -0.4161  2.2555  0.9020 -0.6049 -0.5600  0.3659  0.4109 -0.7661
#> -0.3807 -0.5661  1.0487  1.4307  0.3609 -0.8674  2.2107 -0.8706 -0.7160 -0.5798
#>  0.4461 -1.0077 -0.8379  1.0716 -1.3132  0.6236 -0.0674  0.3987 -0.9534  0.8583
#>  0.5177 -1.1336  1.9239  0.0955  0.7517  0.9753  0.5773 -0.8466  0.5346 -0.7057
#>  0.1700  0.2399 -1.0759  0.4659  1.1433  0.9916  0.5502  0.7893 -0.8221 -0.5467
#> -0.0567 -0.0761  0.4162 -0.0736 -0.1284 -0.4423 -0.3921 -0.7780  0.5004  0.2259
#> -0.9059  0.0377  0.5479  0.4387 -0.2002 -1.1399  0.7167  0.4086  1.1358 -2.0622
#>  0.3160  0.3278  0.4600  0.9376 -0.5667  1.5536 -1.1898 -1.6654 -2.0755 -1.0800
#> -0.3528  0.4002 -0.2552 -0.6200  0.1175  3.5155 -1.3824 -0.2724  1.7088 -1.6368
#>  0.0333 -0.4146 -0.8441  0.3479 -0.4718  1.6045 -0.0726 -0.5174  1.1506 -1.4113
#> -0.0245 -0.9827 -0.6983 -0.8485 -0.5684 -0.4472 -1.1273 -0.7227  0.6406  0.6701
#> -1.4048 -0.0558  0.0350 -0.0478  0.0859  1.5986  2.0529  0.3402  0.0962 -0.4256
#>  0.8660  0.5204  0.8029 -0.4237 -0.7479  0.6550 -1.0743 -0.3253 -0.8245  0.5786
#>  0.8675 -1.7220  0.4298  0.1082 -0.4043  0.6103 -0.0227  1.8835 -0.2437  1.8222
#>  0.8737 -0.8792 -1.1509  0.4871  0.6008 -0.3003  0.3920  0.1563 -1.0617 -0.1801
#>  0.5700  0.1526  1.0177  0.5543  0.2172 -2.3781  0.0230  0.6074 -0.8977 -0.9221
#>  1.8445  0.8246 -0.1957  0.0331  0.7262 -2.0581 -1.6555 -2.2741 -0.1840 -0.0754
#>  1.3179 -1.1248 -2.1237  0.9210  0.1723 -0.1879  1.6754  1.1844  0.1070  0.5615
#> -1.5418 -1.4397  0.2312 -1.5931 -0.4825  0.0574 -0.8901  0.7758 -0.4446  0.3810
#>  1.0759  1.4366  1.3087  1.4185  0.5913  1.3403  0.1260 -1.1927  1.2808 -0.4835
#> -0.3123 -0.7477  1.1050  0.4010 -0.8514 -0.2800 -2.1699  0.0922 -0.2192  0.2402
#> -1.3928 -0.2752  0.2885  0.9339 -0.8284 -0.0150 -0.0685 -1.2067  2.0285  1.3233
#>  1.7209  0.6099 -0.6301 -0.4984 -0.4423  2.0453  1.2661 -0.0785  0.8250  0.5204
#> -0.0837 -1.0434 -1.0200  0.8413  0.4105 -1.8607  2.1539  1.9921 -1.2735  1.0337
#>  0.2401 -0.1198  0.2817 -0.2903 -0.0134 -1.1193  0.0233 -1.6299 -0.2526  0.7240
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]