Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4479  0.6945  0.3696 -0.6583  0.0973  0.3180  0.3268  1.3360 -0.6442 -1.5203
#> -0.7311  0.5139  0.0269 -0.5545  0.4832  0.2656 -0.2673  0.3620  0.0922 -0.5932
#>  0.8532  2.2665 -0.6828  0.5125  0.1034 -0.9357 -0.1550 -0.7360  0.6679 -0.9356
#> -0.1874  1.5815  1.1173 -0.4983  0.0570 -0.6218  0.1082 -0.1103 -0.2736  0.6158
#>  2.6168  0.7586 -1.5046  1.6602  0.3081  0.3213  0.9699  0.3444  0.6684  0.4231
#> -0.9474  0.0962  0.5370  0.3981  0.4160 -0.4767  0.3969  0.9410  0.2801 -1.2378
#> -0.3261 -0.7352  1.0739  1.1779  0.0439 -0.1762  1.5504 -1.3910 -0.0056  0.5494
#> -0.4382  0.3527 -0.3578 -1.2609  0.8177 -0.7608  1.7915  0.9890 -0.3632 -1.6767
#>  0.0274 -0.4485  1.0889 -1.1957  0.2758 -1.2051  0.2524  0.1332 -0.0809  0.2244
#>  0.9276  2.1356 -0.1122 -0.5215  0.5357  2.3860  0.0439  0.5346 -0.2774 -1.7301
#>  1.3439 -0.8727 -0.9466 -0.8197 -1.9300  0.2231 -0.4116  0.3209 -0.2947  0.1242
#> -0.2862  0.3392  1.3970 -1.5847  0.5717 -1.0653  1.2929 -1.4083 -0.0766  0.0261
#> -0.1947 -0.0121 -0.6135 -2.1065  2.9076  0.8891  0.6664 -0.6920  1.4400 -0.8287
#>  0.5407  1.4095  0.4504  0.8664  0.9853  0.7738  0.9121  0.3315 -0.1118  1.9165
#> -2.1454 -1.9085  0.0636  1.2703 -0.4357 -0.2536 -0.6319  0.2055 -0.8036 -0.6227
#> -1.0189 -0.5266  0.0575  2.9518  1.1078  0.0691  0.2140  0.4657  1.0825 -0.6037
#>  0.7016 -0.2174  0.5272 -1.0719 -1.1120  0.1609  0.1782 -0.3234  0.1218 -2.2553
#> -0.2224 -1.1191 -2.4580  0.6178 -0.6153  0.6326  0.6109 -0.1904  0.5060  0.6356
#> -0.8845  1.0491  0.7787 -1.0756  2.3160 -0.6092  0.4135  0.0574 -0.4839 -0.0343
#>  0.0052 -0.3716  2.1986  0.2106  1.1712  0.5910  0.8400 -0.5684  1.7216 -1.3276
#>  1.2017 -1.3632 -1.4929 -0.3705  0.2251 -0.4223 -1.8472 -0.1013 -0.9571  0.3093
#> -1.7320  1.2922  1.4384  0.8179  1.3083  0.7808 -0.4682  0.1657  1.1116  0.9441
#> -0.3643 -0.3699  1.0205 -0.6178 -0.7593  1.9312  0.9009  0.3207 -0.3011 -0.4145
#> -1.1205 -0.4568 -0.8880 -0.0972  0.2838 -0.7912  0.0089 -0.5957 -0.3329 -0.1466
#>  1.3851 -0.7944  0.7703 -0.8213  0.4177 -0.1578  0.4518  1.3370  1.7894 -0.1479
#> -0.7933 -0.8749 -1.1852 -1.1038 -1.1075  0.9352  0.7095  1.4103  1.7990 -1.1115
#> -0.7688 -0.3473  0.3978  0.3620  0.2465 -0.2289 -0.1234  0.2714 -1.2640 -0.6625
#> -0.6324  2.0944 -0.2412  0.9734  2.3681 -0.1368 -0.0981  0.5621 -1.0734  1.0941
#> -0.9060 -1.9821  2.2090 -1.4653 -0.4507  0.2179  0.6762 -0.7560  0.0894 -0.2478
#>  0.9414  0.8289 -2.7575 -0.5276  0.6027 -0.3122  0.5327  0.1117 -1.4792  0.4235
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]