Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.2876  0.2078 -1.4653  0.2176  0.1061  0.8340 -0.6688 -0.6649 -1.7579 -0.4235
#>  1.4246 -1.5327 -0.2620 -1.2579  0.4340 -1.3917  1.0763  0.5040 -0.7974 -0.2960
#> -0.4100 -0.8506  0.7705  0.5731 -1.8080 -1.4407 -0.2812  1.3103  1.3856 -0.1345
#>  0.7735  1.3325 -0.0235 -2.1902  1.1468 -0.4776 -1.5901 -0.1453 -0.3769  0.0839
#>  2.0086 -0.3959  0.6424 -0.0372  1.1386  0.8827 -0.9441  0.2911  0.7506  1.3965
#> -0.0151 -0.3898 -0.4955 -0.1621  1.5077  1.3245 -1.4618  1.6408 -1.2876  0.2947
#>  2.1302  2.6491 -0.3969  1.1449  0.2416  1.3159 -2.3712 -1.3510  0.9878  1.7183
#>  0.3273 -0.9461  0.0224 -0.3363  1.5052  1.3159 -0.6292  0.0406  1.5401 -0.2807
#>  0.2880  0.8493  0.1409  1.0969 -0.3728  0.9905  0.2276  0.2923 -1.4037 -0.4223
#> -0.7444 -0.7564 -1.3047  1.8980  1.1391  0.2406 -1.4845  0.6081  0.2220 -0.2756
#>  0.1249  0.1558  0.1396 -0.0140  0.2749  0.3815 -0.3024 -0.0575 -0.2123 -0.3799
#>  1.4172 -1.0531 -1.0852 -0.6068  0.3270 -1.0306  0.2191  1.4988  0.9132 -0.0481
#>  0.1564 -0.2467 -0.8496 -0.1118  0.1599  0.2180 -0.1123  0.6514  1.1189  0.5639
#>  0.2381  0.7529  0.4471 -1.6046 -0.1795 -0.3365  2.7627  1.5158  0.1449 -1.2906
#> -1.0813  0.6601  0.9644 -0.7498  0.7373 -0.9581  0.7724 -1.0408 -2.7455 -0.4205
#> -0.8072  1.1282  0.4770  1.8003 -0.3572  0.7701 -0.8480  1.1337  1.0869 -0.1532
#>  0.8351 -0.2303  0.3588 -1.0061  0.3961 -0.5042 -0.3978  1.0506  0.3085  2.3232
#>  0.1044  0.2956 -1.2035 -0.7879  1.2680  0.9898  0.2257 -2.2203  0.5179  0.3859
#>  0.7643  0.2608 -0.5795 -1.5269  0.3846 -0.1272  0.5076  0.8798 -0.4068 -0.8685
#> -0.8041 -0.4263 -1.0997  0.2687  1.2278  1.5982  0.1541 -0.3793 -0.6444  1.9063
#> -0.1319 -0.9282 -0.7090 -0.4209 -0.0701  0.5517  0.0054  0.4676 -1.0415 -1.6123
#> -0.6919  1.2040 -0.3401  1.7997 -0.2889 -0.3291  0.6165 -0.3077  0.4007  0.7757
#>  1.5897  0.9689  0.7113  0.7762  0.1475 -0.3797 -1.0196  1.0802 -0.3612  0.6645
#> -1.6886  0.7864 -1.1144  0.6336  0.7595  0.4008  0.0987  1.2731  1.7346 -0.9805
#> -0.0338  0.2009  0.7745  0.8523  0.8979 -1.4443 -0.1191  1.3536  1.6739  0.3782
#> -1.0295 -2.4035 -0.3202  0.2817  0.7878  0.0430  0.1940 -0.9713  0.6808  0.9558
#> -0.1002 -1.3419 -1.0261 -0.9142 -1.0406 -0.0052 -0.7415 -0.5210  1.8885  1.0857
#>  1.3535  0.8646  0.6405  0.9371  0.7271 -0.2145  0.5364 -0.3508  0.2482  0.1225
#>  0.2649 -0.8389  0.1545 -1.2856 -1.5248 -0.6197 -0.6808 -1.3653  0.7566  0.1116
#> -0.1407 -0.1048  2.7640  0.6351  0.4356 -0.9332  0.1164 -1.0940  1.3109  0.7791
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]