Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 2.7620e-01  2.1638e-01  5.7402e-01  1.3348e+00  2.6626e-01 -1.1868e+00
#> -6.3026e-01  1.2956e-01 -9.9690e-01  3.5070e-01 -7.0201e-01  9.5662e-01
#>  2.4882e-01  8.9006e-01 -4.9182e-01 -2.3017e-01 -1.1697e+00 -3.5301e-01
#>  1.1638e+00 -7.6727e-01  1.0118e+00  1.0265e+00  2.0452e+00 -6.3736e-01
#>  6.9046e-01  7.3668e-01 -7.5562e-01  2.0306e-01 -1.4703e+00  3.1863e-01
#>  1.1587e+00 -1.3888e+00  1.1252e+00  2.1359e+00  1.7661e+00  5.0678e-01
#>  1.4713e-01 -5.6429e-01  8.3041e-01 -1.1528e+00 -6.9583e-02  4.5774e-01
#>  1.0112e+00 -7.7139e-01 -7.1905e-01 -1.9552e-01 -1.5181e+00 -1.7682e+00
#>  1.3850e-01  1.0291e-03  7.8814e-03 -9.2687e-02 -4.2570e-02  1.3739e+00
#>  2.5507e-03 -7.9017e-01 -5.5776e-01  7.5583e-01  9.2745e-02  7.2291e-01
#>  9.1832e-01 -1.4885e+00 -6.0998e-01 -5.4409e-02  6.8827e-02  8.5069e-02
#> -3.4793e-01 -5.1098e-01  2.0277e-01 -1.4823e-02 -4.7850e-01  2.1001e-01
#> -1.0302e+00  3.3433e-01  6.3423e-01  7.3425e-01  3.1575e-01 -6.5539e-01
#>  2.4728e+00 -1.1260e+00  4.2090e-01  1.5329e+00 -1.0251e-01  1.0675e+00
#>  7.4068e-01  3.8461e-01  9.5240e-01 -5.3347e-01 -1.8713e+00 -1.2545e+00
#>  7.5099e-01 -8.4092e-01  2.0236e-01  3.7753e-01  1.6856e+00  2.7415e-01
#>  6.5410e-01 -2.4555e+00 -5.0759e-01 -7.4651e-01 -1.1968e+00  6.0663e-01
#>  1.3931e+00 -9.1235e-01 -2.4148e-01 -1.0050e+00  1.1500e+00 -1.6773e+00
#>  7.8959e-01  9.2283e-01 -2.2794e-01  2.1806e+00 -1.5816e+00  9.6108e-01
#> -2.7474e-01 -1.4548e-01 -4.1815e-02 -2.1581e+00 -1.1679e+00 -4.5480e-01
#>  1.2149e+00  9.6653e-01  2.0184e-01 -1.2012e+00  1.6785e+00 -1.3767e+00
#>  3.9642e-01 -7.2946e-01  5.3119e-01  1.1963e-01 -1.0440e-01  8.5725e-02
#> -1.6758e-01 -7.5238e-01 -1.1322e+00  8.8371e-01  1.4849e+00  5.2717e-01
#> -5.1076e-01  5.0616e-01 -5.2045e-01 -6.5247e-02 -2.2276e+00  4.5980e-01
#> -1.8922e+00  7.1640e-01 -1.0250e+00  9.0236e-02 -6.8276e-04  7.4353e-01
#>  9.9064e-01 -1.5877e-01 -1.2909e+00 -7.9439e-01  4.4436e-02 -5.3647e-01
#> -4.5757e-01  2.8181e-01  7.0978e-01  5.7012e-01 -6.9011e-01  8.6969e-01
#> -7.8672e-01  3.9294e-01 -1.0618e-01 -8.7267e-01 -5.6734e-01  2.1976e-01
#> -2.4628e-01  1.2145e+00 -2.1675e+00  3.9551e-01 -1.4252e-02 -2.0240e-01
#> -2.7021e+00 -1.0288e-01  1.1170e+00 -3.1014e-01 -1.1925e+00 -5.9297e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]