Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0052  0.0593 -1.0688 -0.7872 -0.5812  0.7913  0.5330 -1.1805 -0.4956 -0.7607
#> -1.1389  1.2794  0.6748 -1.8505 -0.7708  0.4723  0.5917  0.6871  2.1063  1.0737
#>  0.4516 -0.7568 -0.1488 -1.5378 -0.1193 -0.3665  0.0234 -0.9306  2.4411 -0.6763
#> -0.5304 -0.5948  1.6949  1.3900 -1.5722  0.4567  0.4985  2.1066 -0.6295  0.1500
#>  0.5279  0.2869 -0.2553  1.2105 -1.4248  0.3143 -1.8187 -0.6152 -0.7143  0.3884
#> -0.6191 -0.1012  0.4086 -0.5219  0.6506 -0.8191 -0.2525  1.1599  1.5180  0.4237
#> -1.0538 -1.8513  0.6350 -0.7614 -0.2484  0.1268 -0.4122  1.0237  1.5164 -1.1413
#>  1.5045 -0.5384  0.0311  0.1425  0.2606  1.8986 -0.7881  1.3461 -0.8838  0.0582
#>  0.4669 -0.3599 -0.7377  0.3402  1.8225  0.6123  0.4711 -0.7374  0.3853  0.7262
#>  0.0981  0.6165  1.4381  0.6228 -2.1716  0.1975 -0.7608  0.0816 -0.7715  0.2450
#> -0.6418  0.4013  0.0119 -0.1663 -0.6126  2.1559 -1.0957  1.0734  0.0837  0.4675
#>  1.4594  0.4077 -1.4502 -0.7905  0.3029  0.9517  0.2698  0.2615  0.2460  0.0203
#> -1.0618 -2.1162 -1.1310  0.1777 -0.2133 -2.2476  1.3278 -1.7698 -1.3400 -1.2217
#> -0.8028  0.8289  0.4166  2.1225  0.9257  1.0167 -0.6522  0.3357 -0.0234 -0.6523
#> -0.0651  1.2597 -0.5609 -0.9037 -0.4464 -0.8845 -1.0018 -0.8431  1.0042  2.7845
#>  0.5500  0.4931 -0.6640  0.3646 -0.8212 -0.3195  0.1352 -0.3488  0.5602  1.3482
#>  0.5870 -0.4875  0.9029 -1.6088 -0.7369  0.7552  1.0988  0.3868 -0.0627 -0.5273
#> -0.0206  0.4331  0.8740 -1.2247 -1.3441 -0.6518  2.1823 -0.4657  1.3370 -1.1892
#>  0.1704  0.5649  1.2863  1.6334 -0.9626  0.1246  0.2274 -1.0967  0.3471 -0.1259
#> -1.0618 -1.9489  0.8544 -0.0652 -1.0664 -0.3805 -0.4783  0.9574  0.3624 -0.7733
#> -0.4222 -1.1841 -0.1307 -1.1649  0.9502 -1.8153 -0.5203 -0.6300 -0.0462 -1.4746
#> -0.1019 -1.5217  0.2652 -0.2003  0.4736 -1.7079 -1.1007 -0.6885  1.9725  1.2034
#>  0.8824 -0.1872 -0.8122 -1.0760 -1.9483  0.7866  0.7785  0.9796  0.7054  1.1037
#> -0.1631  1.0624  1.8601  0.4705  0.5344  0.9246 -0.4019 -1.0514  0.0204 -0.1038
#> -1.5963 -0.6082 -1.7366 -1.7006 -0.5045 -0.4957 -0.0338  0.0920  0.2466 -0.3115
#>  1.1813  0.8559  1.3089  0.1631  0.3597  0.1222 -0.3503 -0.3906  2.1412 -1.0152
#> -0.5189 -1.5532 -0.8010  1.4698 -0.9119  1.7737 -0.6634  1.4085  1.0274  0.2893
#> -1.6838  0.6860 -0.3138 -0.9018 -0.6502  0.9962 -0.0112  0.0465  0.6851 -1.2565
#>  1.4657 -1.3323  1.9344 -0.5363 -0.3826 -0.8723  0.3267 -1.0460  1.5912  1.1643
#> -0.4947 -0.2933  0.8451 -2.4080  0.1228 -0.5445  1.8050 -0.6418  0.0743 -0.3859
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]