Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.7542 -1.0078 -0.7417 -1.0975 -0.3025  1.3496  1.2671 -0.7937  0.0585  1.3373
#> -1.4084  1.0885 -0.2891 -0.2188 -1.9558 -0.7206  0.1502  0.5948  1.1537  2.0775
#> -0.2084  0.5449 -0.1560  0.8568  1.5726 -0.6966  0.6029  1.5145  0.3653  1.2384
#> -0.0679  0.6201  0.0267  0.2960 -0.5715  0.0763  0.8008 -1.2117 -1.4284 -0.3447
#>  0.2390  0.7964 -1.3983  0.6099  0.3755  0.5884 -0.1492  1.1194  0.2763 -0.4827
#>  0.2772 -0.6744 -1.0824 -0.0857  0.3074 -0.4439 -1.4928 -0.8293 -1.6158 -0.0314
#> -0.7675  1.1862 -1.0241 -0.5584  1.5580  0.3028 -0.1598 -1.0371 -0.7810  1.1833
#> -1.3651  1.6606 -0.6793 -0.8551  0.5221 -0.0793  1.2946  1.5098  0.1302  0.1737
#>  1.7288 -0.1836 -0.5128 -1.1196 -0.4898 -0.4403 -2.0523  0.8025  0.2542  0.8161
#> -0.2257  2.2978  1.0028 -0.4877  0.5175  0.3791  3.2846 -0.7002  0.2696  0.7908
#>  0.4851  0.5883 -1.3735 -1.0102  0.3533 -0.2588  1.6209 -0.6720  2.2927 -1.6944
#>  0.0315 -1.9618 -0.3414  1.3377  0.4730 -0.5635  0.2120  0.9639  0.4115  0.6639
#>  0.0692 -1.4218  1.6734 -0.2725  0.6420 -0.2140 -2.2127 -0.3421  0.1854  0.6737
#> -0.5899 -0.8794  1.8207 -0.6163 -0.8787 -0.1808  1.6749  0.8491 -1.3940 -1.6242
#>  0.6108 -1.0678 -0.5422 -0.8071  0.1616 -0.7216  0.3433  0.3670 -1.8974 -1.3646
#>  0.5218 -0.8242  0.7188  0.0226  1.7397  0.2942  0.2231 -0.3259  0.5256  0.7498
#> -0.5336  1.5268 -0.8935 -1.2531 -1.4541  0.6142  0.8101 -1.3845  0.8956  1.3284
#> -1.2136 -1.1427 -0.3197  0.5047 -0.4325 -0.5044 -0.3435  2.7859  0.4804  0.1884
#>  0.1363 -0.4961 -1.2338 -0.5652  1.5842 -0.4594 -0.4859  0.0785  1.0483 -0.1833
#>  0.0872 -0.7250  0.9126 -1.8285  0.4393 -1.3183  0.3612 -1.8864 -0.0668 -1.2429
#> -0.5642  0.5388  1.7785 -0.3869 -0.3992 -0.2081 -0.3241  2.5055 -1.4425  0.6781
#>  0.0697  0.3912  0.8506  1.8309  0.9621 -0.3807  0.5005 -0.5517 -0.8954  1.1762
#>  0.3021  0.1362  1.4594 -0.0101  1.1564  0.7725 -2.2655  0.9197 -1.7619  0.3077
#> -3.0552  1.0027  0.5944  1.0487 -1.3328 -1.6735 -0.2445  1.3449  1.0861  0.4655
#>  0.0594  0.7266 -1.6100  0.2421  0.0550  0.2531  0.5711  0.9646  0.0211 -0.1360
#>  1.6660 -0.3760  0.1254  0.6772 -0.1125 -2.2012 -0.0025  0.2783  1.2710  0.3177
#> -0.1146 -2.6522 -0.7854 -1.7416 -0.1659  0.2448 -0.9926 -1.3746  0.1265  1.9924
#> -0.1509 -0.1797 -0.6086  3.4691  0.4170  0.2684 -1.9153  1.8793 -0.6754 -0.4563
#>  1.1991 -0.1699 -0.9000 -0.2574 -1.9088  0.7593 -0.7001  0.1162 -0.6962  0.8413
#> -1.1168  0.4147 -1.9708  1.5555  1.8505 -0.5866  0.9663  0.1667  1.1177  0.0576
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]