Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.2631 -1.2414 -1.8945 -0.0832  0.6123  0.0417 -0.9687  1.1433 -0.6567 -1.5487
#> -1.2380  0.5978 -0.6813 -1.0301 -0.1184 -0.0821 -0.5029  0.3566  0.0900 -0.3744
#> -0.9676 -0.7595  0.7710 -1.4004 -1.2945  0.8719  1.1100  0.2703  0.2572 -1.1684
#>  0.0507 -1.8151 -0.9152 -0.8035 -0.2657 -2.0176  1.0038 -1.8281 -1.0301  2.1503
#> -2.0700 -1.5323 -0.6261  0.8150 -2.0028  1.3269  0.9649  0.2745  0.3312 -0.1887
#>  0.1931 -1.1457  0.4022  1.2606 -0.8179 -1.3634  0.3004  0.3965 -1.1222  0.8231
#>  0.1740  1.2821 -1.2667 -1.0671  0.6520  1.7805  1.6986 -0.4028 -0.5077 -1.7189
#>  1.1447  0.0532  1.5603  0.6457  1.7200 -0.7450  0.3805  0.5083  1.5593 -0.7414
#>  0.4390 -0.1425  0.0188 -0.3161 -0.7030  0.0884  0.6006 -0.6623  0.7676  0.6334
#>  1.2207 -0.5597  0.7563  0.4049  0.8365 -0.2074  1.0292 -0.7654  1.9295  0.9602
#> -0.0082  0.1599 -1.2730  2.1090 -1.8153 -1.6925 -0.4525  0.2376 -0.7994  0.3705
#> -0.3199 -0.6910 -0.5644 -1.8129  1.0302 -0.8992 -0.3766  1.1428  0.8892  0.4840
#> -1.2091 -0.1677 -1.2241 -0.1017 -1.9259  0.2685  1.6889  0.2204  1.8693 -0.3479
#> -0.3548 -1.0136 -0.5211 -0.0455 -0.0531  0.4348 -0.8736  0.4743  1.6573  0.0620
#> -1.1968  0.8317  0.8749  0.6987  0.1670  0.8204  0.1864  0.6088 -1.2820 -1.0808
#> -0.2188 -0.7130 -1.6688 -1.9214  1.2715 -0.4349 -0.5589  0.5130 -0.1197 -0.2663
#> -1.2120  2.2128 -0.8050 -1.3269  0.3350 -0.7813 -2.1490 -0.3231  0.3030  1.3257
#>  0.7070 -0.5038 -0.8938  0.4148  0.0039  1.3793  0.6356  1.0763  1.3372  0.0403
#> -0.2738  0.5955 -0.4092 -0.2321 -0.9272 -0.5713 -1.1944  1.0524 -1.0804 -1.0506
#> -0.2833 -0.3009 -0.0856 -0.0251 -0.2038 -1.0525  1.0600 -0.4886  1.7715  0.6450
#> -1.0049  0.3614  1.1893  3.0472  0.9601  1.2316  0.2120  1.0503 -0.1828  1.1644
#> -0.8248  2.0644 -1.5502  0.0749 -0.6771 -0.3550 -0.4892  1.3529 -0.3461  0.5551
#>  1.0789  0.7013  0.8025  0.2676 -2.1081 -0.9392 -0.3319 -0.8354  0.6374  1.5051
#> -0.4433 -0.4891 -0.1209  0.5542 -0.1500 -0.5630  1.4630  0.5518  0.2721  0.0936
#>  0.7817  1.0698  1.4646 -0.0164 -0.9314  1.1440 -0.5521  0.2700  0.3075  1.4460
#>  0.6915 -0.0281  0.6617 -0.4319 -0.2237 -1.6404 -1.1335 -0.3147  0.2470  2.7856
#> -2.2841 -1.0548 -0.2148 -0.2609 -0.5116 -0.7766 -0.2459  0.3459  1.8629 -0.5950
#>  1.7940 -1.7374 -0.2227  2.0236 -0.4777  1.2078  0.9016 -0.1997  2.1726  0.4732
#> -1.5126  0.3140  0.9492 -1.7454  0.5103 -0.4529 -0.6944  0.5154  0.4328 -1.1285
#> -0.3450 -0.3852 -1.3898  0.7118 -0.2136  1.3462  0.1002  1.2430 -0.6810 -1.5397
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]