Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.7889  0.0160 -0.9466  1.9892  1.0669  1.2490  0.9614 -0.6404  0.4553  0.6799
#> -0.8546 -1.0591 -1.7655  0.6600 -0.8366  0.3642  0.6432 -2.6365  0.4407  1.3924
#> -0.0311  0.4841  2.0319  0.8586 -0.5361 -0.1059  0.4666  0.3552  0.3429 -0.2947
#> -0.0485  1.6201  1.2192 -0.6417  0.5167 -0.2638  0.7654 -1.6884  0.3372  1.3074
#> -0.8856  0.3237  0.0248 -0.1735 -0.5903 -0.8658  0.6102  0.4717  0.4893  0.7068
#> -0.5856  0.6528  0.4286  0.4406 -0.7278  0.9373 -0.4161  0.9027 -0.4118 -0.9590
#>  0.1913 -0.4542  0.1339  0.9625  2.5651 -1.3289 -0.0608 -3.3103  0.0906 -0.8095
#> -2.5895 -0.0685  0.1355  0.8576 -2.1713  0.6645  1.4099 -1.0230  1.3932  2.3931
#>  2.0386  1.4635  0.0607  1.8397  1.6443 -1.2350  0.0702 -1.6000 -1.5556  0.9102
#>  0.2157  1.3477  2.4016 -0.2738  0.2422  0.4543 -0.2058  0.8307 -1.9767  1.2985
#> -0.1396  0.0218  0.3911 -0.5886  1.2025  1.6724  1.4104 -2.2506 -1.2917  0.9555
#> -0.5919 -0.8731  0.7999 -1.1762 -0.4134 -1.7451 -0.2535 -0.8374 -0.1556  0.9269
#> -0.1323  0.8212  2.7056  0.3186  1.0146  2.6063  0.5384  1.5678 -0.0766 -0.1010
#>  1.6156 -0.7140 -0.6422  1.5843  1.4199  0.0489  2.3296 -0.0734 -0.7888  0.7994
#> -0.4659  0.4021 -0.7800 -0.4884  0.1648 -1.9028 -0.4843  0.9192 -1.5620 -1.4710
#> -0.8921  1.3588 -0.2365  1.4491  0.2769  0.2540  0.8105  0.5504 -0.5694 -1.0611
#> -0.9689 -0.8392  0.4950 -0.0507  0.4482 -0.0886  0.6813 -0.6873 -0.8544  1.8588
#>  0.8660  1.1319 -1.4257 -0.4972  0.1747 -0.4053  0.4925 -0.2399  0.5633 -0.8958
#>  0.4350  0.4937 -1.4868  1.4921  0.0234 -1.9182  0.1272 -0.3692  0.4496  0.7617
#>  0.1111 -2.3290 -1.7129  0.6156 -0.4156  0.9154 -1.4722 -1.0123  0.8288  0.5251
#>  0.4268 -0.1959  0.6551 -0.8262  0.2111  0.0291 -0.4715  0.6942 -1.2752 -0.9702
#> -0.9684  1.8245  0.0186 -0.7523  0.0392  0.5005  0.6071 -1.4846 -2.5213  0.6375
#> -0.9757 -1.1397 -1.4060 -0.4228 -1.0215 -2.1740 -1.6158 -0.9298 -1.2088  1.3859
#> -0.4120 -1.0499  0.0995 -0.9767 -0.9448  0.3056 -0.6664 -1.5148  0.5080  0.0041
#>  0.8186 -0.3876  0.0122 -0.7527  0.9305 -1.7900 -0.3935  0.0855 -1.7824  0.7759
#> -1.1903 -0.1806 -1.3260 -0.6132 -0.3847  0.5300 -0.9429 -1.1921 -0.5260  1.6669
#> -0.4267 -0.4178 -0.1257  0.0801  0.4086 -0.1047  1.4768 -2.1072  0.0865 -0.7660
#> -1.1559 -1.3133 -1.1350  0.2537 -0.6192  0.5633 -0.6628  1.7296 -0.3344 -0.7392
#>  0.4645  0.7649  0.6799  1.0866  1.0529  1.0006 -0.4245  0.5042 -0.3410 -0.6732
#>  0.1615 -0.0715  0.5592  1.3820  0.0423 -0.4354  0.5708  0.6695  0.8562 -0.2299
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]