Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.3749  0.1596  0.9282  1.6222  1.1073 -1.3600 -0.9398 -0.0797 -1.4957  0.4255
#>  0.1340 -0.6694 -0.8127 -0.8616 -0.6991  1.1683  0.8721  0.4539  1.2939  0.6770
#> -0.6031  0.9414 -0.0866 -0.4987 -0.6227  0.6253 -0.7833 -1.2599 -0.5639 -0.3695
#> -0.7933  1.0266 -0.0230 -0.4824 -0.3533  1.2552  0.0619  0.5764 -1.7845  1.2304
#> -0.1416 -1.4268  0.9088  0.7837 -0.7257 -0.5577 -1.5208  0.8086  1.3730 -0.0476
#>  1.5874 -0.1804  0.7522  0.2734 -3.0458 -0.6011 -0.3625  1.2731  1.3420  0.1231
#>  1.0847  0.5707 -0.0331 -0.4064  1.4149 -0.6797 -0.8338  1.0807 -0.6364 -1.8007
#> -0.3319 -0.9815 -0.0199  1.1768 -0.4984 -0.3010 -0.8645  0.1579  0.7676 -1.0782
#>  0.2125 -0.2280 -0.7628 -1.7777 -0.1437 -1.9467 -0.3036  0.2929  1.5809 -1.3649
#> -0.3909 -0.6657  2.3331  0.2060 -0.0986  0.8406  0.6784 -0.0016  0.8760  1.6802
#> -2.1857  0.2854  0.0464  0.8140 -0.4859  1.5800 -0.6153 -0.2065  0.2741  0.2783
#> -0.2189 -1.2756 -0.3431 -2.1660  1.0249  0.6195  0.4411  0.2440  0.8722 -0.3350
#> -0.5662 -0.5135 -0.4469  0.3143  0.2811  0.1706 -1.4646 -0.4270  1.3939  0.4988
#>  0.1301  2.2834 -0.2546 -0.1776 -0.7585  0.4357 -1.7997 -1.0512  0.4877 -0.7650
#>  0.5708 -1.6862 -0.6916 -1.7721  1.1633  1.3531 -0.9038  0.8418  1.8312 -0.4681
#> -0.0473  0.9124 -1.8123 -0.0772 -0.4991 -0.5903 -0.9043 -0.8421  0.8211  0.7297
#>  1.1643  1.2940  1.3685  0.1733 -0.4682  2.0195 -0.5432  1.0499  0.2074 -1.7402
#> -0.5053  0.9295 -1.5670  0.4075 -2.4381  0.9241  0.5927  0.2387  0.5661  0.5338
#>  1.4042  1.0360  0.5392 -1.6991  0.0206 -1.3823 -1.2059 -0.0183 -1.4034 -1.2446
#>  0.5750  0.1691  0.4931 -0.3412  0.3971 -0.8876  0.4025  0.1253  0.9968  0.4146
#>  0.4068  1.7906 -0.4032  2.1213  0.2069  0.4632 -0.2407  0.7347 -0.7955 -1.4865
#>  0.5629 -0.3389  1.1529 -0.2400  0.0792  0.3780  0.0025 -1.9673 -1.0596 -0.6208
#>  0.8865  0.0229 -1.1632 -1.2466 -0.9965  0.2321 -0.1061  0.8796  0.7556 -0.3941
#>  0.7572 -1.6926 -0.4047  1.9066  0.4785  1.5193  0.8137 -0.7262  0.8455 -1.1189
#> -0.5737 -0.4700  0.6368  1.2564  1.1105 -0.1861 -0.0250 -0.4801  1.3613  0.8253
#>  0.0435  2.1481  0.3785 -0.1553 -1.3903  1.2150 -1.9697  0.5108  0.7564 -0.2593
#> -0.5857  0.3320 -0.6243  0.1723 -0.8255 -1.6525 -1.6261 -0.0391 -0.7078 -1.3433
#> -0.7324  0.1545  0.1687  0.0636  0.0456 -0.6010 -1.0462 -0.9240 -1.2945 -0.4532
#> -0.7587  0.4712  1.0227 -1.8167  1.7015  0.8298 -0.8796  1.7614 -0.7823  0.2943
#>  0.6575  1.8811 -0.2927  0.8790 -2.1079  0.3034 -0.8527  1.1957  0.7529  0.0774
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]