Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.3407 -0.6894 -0.4665 -0.9412  1.2239  0.4915  0.4589 -0.1204  0.5725  1.2204
#>  0.1290 -0.5753  0.3541  0.0566  1.3324 -1.0410  0.4132  0.7985 -0.8166 -0.0295
#>  0.6125  0.2516 -0.7409 -2.0528 -0.7030  0.8972  0.2683 -0.0985  0.9960 -2.3329
#>  0.9398  1.4511 -0.4682  0.2242  0.3234 -0.9579  0.5432 -2.0606 -1.2099  0.2810
#>  2.0229 -0.4927  1.8863  0.0909  1.6997  2.1289  0.6407  0.5339 -0.1677 -0.7021
#> -0.4315  0.5687 -1.1320  0.2139  0.1582  0.5955 -2.1404 -0.0546 -1.6632 -0.0674
#>  1.7382 -0.6420 -0.7310  0.4235 -0.1522 -0.4887 -0.7671 -0.4688  0.3707  0.8228
#> -0.0892 -0.5136 -0.1966 -0.9877  0.3428 -0.1041  0.0521  1.1204 -0.8384 -1.5735
#> -1.3654  0.2579 -0.1448 -1.2137 -0.4111 -0.3798 -0.9105  0.3558  0.4400  0.4371
#> -0.0960  0.1319 -0.1740  0.9267 -0.0680 -0.4555  0.7220 -0.7796 -0.7670 -0.9805
#>  0.0095  0.5867 -1.3713  1.3365 -0.2331 -0.2839 -0.5833  2.7552  1.8778 -1.8715
#> -0.4755 -1.1962 -0.8491  0.3717  0.4049 -1.0132 -0.4768 -0.6612  1.5277  1.6098
#> -0.8100  2.0095 -0.8867  1.9401 -0.0864 -0.3004  0.9664  0.1668 -0.0705 -0.2609
#> -0.1869 -0.5407 -0.6043  1.8052 -0.4040  1.2521  0.7708  1.2560 -0.4789 -0.6612
#>  2.1334  1.5673 -0.8303 -0.6515  0.4236 -2.7794 -0.2529 -0.4103  0.8017  0.3434
#>  1.3768  1.5801  0.4617  0.6819  0.3062  1.1320  0.7167 -0.6234 -0.0507  0.4092
#> -0.2215 -0.3154  1.0567  0.4845 -0.4865 -1.2510 -0.3604 -0.6446  2.7592  0.2948
#> -0.5855 -1.6724 -0.4302  0.9002 -0.2744 -0.6662  0.6390  0.0158 -0.9258  1.3998
#> -0.0769  0.2905 -1.7569  0.0568 -0.6515  0.8263 -0.3350 -0.2899 -0.8312 -0.1760
#>  0.9330 -0.7329  0.6084  0.2609 -0.1955  0.7395  0.5290  0.1627  0.2637 -0.3681
#>  1.7584  0.3046 -1.1893  1.9568 -1.0775 -1.9646  0.4922 -0.9054  1.0295  0.1594
#>  1.6844 -0.9334  0.0345  1.0944  0.2659 -1.9379  0.5225 -0.3455 -0.0088  0.3942
#>  0.5680  0.0373 -0.3285  0.6044 -0.6071  0.0884  0.3219  0.4209 -0.5356 -0.0627
#> -1.1465  0.3588 -1.7775 -1.0463  0.6417  0.3719 -0.6521 -0.2149 -1.2753 -2.0836
#> -0.7034  0.8012 -0.0392  1.8284  0.6644  1.7817  0.2952  0.8439 -0.6417  1.3467
#> -2.3127 -0.0718 -0.5444  0.2454 -0.9047  0.6211 -0.6254 -3.7644 -1.4726  0.3252
#>  0.8890 -1.3176 -0.3843  0.4585  0.8552 -0.6395  0.5924 -1.3633 -0.6235  1.6147
#> -0.7633  0.6676  1.7451  0.3555 -0.3558 -0.8849 -1.1491 -0.0811 -0.6012  2.1837
#>  0.4479  0.1580 -0.1069 -2.4857  1.4795 -0.5301 -0.7482  0.4471 -0.2663 -0.0437
#>  0.5878  2.5901  0.1904  0.3082 -0.6711  0.2174 -1.5495  0.0530 -1.0607  1.1701
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]