Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0014 -0.6146  1.0062  1.0799 -0.3724 -0.4354 -0.5103 -0.3693  0.3527 -0.2329
#> -1.8992  0.0518 -0.6617  0.2457  0.4164 -1.7696 -1.6558 -0.0863 -1.2033  0.2799
#>  0.2310  0.4808  0.9276  1.2224 -0.7757  0.3648 -1.1715 -0.2514 -0.0144 -0.0045
#> -0.9845 -0.8366  1.8990 -0.3206 -0.3326  1.0207 -2.2316 -0.4259 -0.0106 -0.1178
#> -0.6011  0.2593  0.1368 -1.2262 -1.1799  1.0644  0.7908  0.1575 -0.2447 -0.4085
#>  0.1624 -2.1855  1.8025 -0.8446  1.9189 -0.8178  0.9550  0.3483  0.4145  0.5245
#>  0.7978 -0.0497 -2.8168 -1.9326 -0.1984  0.7231 -1.0931 -0.2659  0.8566  1.1357
#> -0.7322  0.1617  0.6915  0.5580 -0.6952 -0.0405 -0.8671  0.3030 -0.2869  0.2974
#> -0.0732 -0.0211  0.4897  1.5216 -1.6484 -0.3528 -0.1556 -0.4576  0.5771  0.3473
#> -0.6944  1.0465  0.3022  0.3842 -2.1889 -1.0498  1.6511  0.1909 -1.4213  0.8304
#>  0.3817 -0.0396  0.6813 -0.4889  1.1144  0.9883  0.6463 -0.9010 -0.7711 -0.4695
#> -0.0244  0.5294  0.3573 -1.5950  0.9399 -0.9372  0.1042 -2.4929  2.4568  0.3855
#> -0.9967  0.4774  2.2189  0.6547  1.4483 -1.1808  0.3644 -0.3729  1.9231  0.7931
#> -0.2554  0.8685 -1.0505 -1.1870 -0.2295 -0.1232  0.6822 -0.8025  0.9479  1.7606
#>  1.2588  1.4957  0.1665 -1.0828  0.5939 -0.5629 -0.5664 -0.6820  0.0249  0.3111
#>  0.0321 -0.0266  0.7256 -0.8200  0.2132  1.7390  1.3487  0.9374 -0.6631  0.0465
#> -0.3488  0.9596  0.2018  0.0810  0.6675 -0.2499  0.0506 -1.4610  1.2138 -0.3551
#>  0.7633 -0.7100 -0.4602 -1.3145  0.4356 -0.0029 -0.2049  2.0101  1.4936 -0.3206
#> -0.5436  0.4143 -1.3987 -0.3709 -0.5167 -0.4823  2.1068 -0.7034 -0.9884  0.8368
#>  0.2110  0.2506 -0.7597 -0.6813 -0.3594  0.9547  0.1643  1.0330 -0.6785 -1.1482
#>  2.0584  0.6236  1.1564  0.5023  0.4773 -1.2131  0.3688  0.2913 -0.3880  0.1703
#>  0.7720  0.0152  1.1058 -0.3296 -0.4650 -0.4333 -1.4460  2.3435  0.4086  0.4680
#>  0.5216 -0.2091 -0.4159 -0.6223 -0.2316  0.9516  0.0360 -1.2137 -0.1126 -0.7890
#>  0.1353 -0.5360 -0.4626  1.8740  0.2256  0.0059  0.8504  0.6165 -1.2352  1.3307
#>  1.3811 -0.5556  1.6518 -0.4528  1.7882  0.9862 -0.1364 -0.3545 -2.7194 -0.8162
#>  0.5039 -0.9925  0.0769  0.3283 -0.4295 -0.5398  0.7497  0.5975 -0.0741  0.0021
#> -0.3815  0.2102 -1.0627  1.2020 -0.3005 -2.0863  0.8317  1.1440 -0.7372 -1.3093
#> -1.1759 -0.7734 -1.7409 -0.8393  0.3974 -0.2412 -1.5825 -1.2581  1.2000  0.9398
#>  0.5202  0.5211  1.4119 -0.3822 -0.9582  1.8024 -0.2655 -0.6224  1.4024  1.5616
#> -0.7421 -1.3647  1.1214  0.1422  1.6892 -0.0484 -1.0841  2.2572  0.4360 -1.1919
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]