Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-1.1372  0.3499  0.0747 -0.3960 -0.6106 -1.3606  0.4125 -0.7725 -0.3790  1.0416
#>  0.3082  0.4275  0.8373  0.7932  1.7294  0.6429 -0.0963 -0.0404 -0.8949  0.1846
#> -0.7096 -0.1454 -0.5794  0.4824  1.4979 -0.0769 -0.5720  1.7564 -0.4428 -1.1787
#>  1.4452  0.8845 -0.2052  0.0120  1.5556 -0.9468 -0.2065 -0.2150 -0.7846  0.3028
#> -0.2338  0.9949  1.7809 -0.0511  0.2465  2.1962 -0.6056 -0.5434  1.5650  1.0291
#> -0.1464  0.5512 -0.4795  1.9145  0.2260 -0.8211 -0.4910 -0.5151 -0.0201 -0.3211
#>  1.3182  2.0178  0.7190 -1.9411  0.4842  0.0149  0.3508  1.0200  1.1169  0.2232
#> -0.2083 -0.2713  0.9713 -1.1857  0.5123  0.3775 -0.8138  2.8327  1.6217 -0.8843
#> -0.2874  0.0786 -0.8818 -0.8508  1.0834  0.5353  0.4201 -0.9509 -0.1690  0.1856
#> -0.1153  1.2933 -0.7881  0.8407 -1.0525 -1.3508 -0.8082 -0.7737  0.8267  0.1241
#>  0.7461 -1.6021  0.0369  0.0607  1.0322  0.0747 -0.6548 -1.8300 -0.8820 -0.9063
#>  0.1667 -0.4978 -0.4302 -0.1489 -0.4490  0.6321 -0.7731  0.8231  0.2661  0.3337
#> -0.7664 -0.8211 -0.4047 -0.2328 -0.3662  1.1360 -1.8555  0.7962 -0.6962  0.0018
#>  0.0890  0.6462  1.9629 -0.4416 -0.9250  1.2053 -1.9468 -0.1588 -0.1932 -0.7440
#>  0.6065 -0.6523 -0.2620  0.0955  0.7917  0.4453 -0.3093 -0.5856  0.9093  1.9465
#>  0.3684  0.1390  0.3845 -1.5951  0.8773 -0.9996  0.2017  0.6218 -1.0841 -0.0838
#>  0.0672 -1.0103 -2.1210  1.2772 -1.6008 -1.8350  1.4602  1.6355  1.1787  0.6150
#> -0.0966 -0.1713  0.8972 -0.3104 -0.9355  0.0967 -0.7840 -2.0124  1.0620 -0.4523
#>  0.5556  0.0227 -2.4423  0.0768  0.7598 -0.5121  0.5880 -1.8360 -0.8892  1.6499
#> -1.4925 -0.8357 -0.0370  1.2093 -0.3215  1.1185  0.2934  0.2208  0.3973 -0.6541
#> -0.8940  0.1878  0.2721  0.4228 -0.8414  0.3170  1.0158 -0.5809  0.5893 -0.1229
#>  1.1276  0.6694 -0.5730 -1.1456 -0.1952  0.0939  1.7643 -0.2802 -1.2060  0.3305
#> -0.9220  0.0134  1.1092 -2.0741 -0.3757  0.7162 -0.0354 -0.8173  1.3979 -1.1718
#>  1.5388  0.6873 -1.0699  0.6491 -0.8015  0.9243 -0.5145  0.1313 -1.1410  0.2603
#> -0.0911 -0.2311  0.8274  0.9381  0.1558 -0.0354  0.5036 -0.3324  0.2745  0.6223
#> -0.0702  0.8225 -1.1387  0.8917 -0.6800 -1.9296  1.0420  1.4478  0.2573 -0.1559
#>  0.2551 -0.4104  0.5489 -0.2696 -1.1426 -0.9160  2.3791  0.0828  1.8574  1.2391
#>  0.5185 -0.3707 -0.1568  0.1843  1.0585  0.7519  0.4969  0.7398 -0.4183  3.0056
#> -1.8073  0.4373 -0.4441  0.4325  1.7536  0.0078 -0.9750 -1.4524 -0.9081  0.6500
#> -1.0542 -1.0404  1.8536  0.1971  0.5177 -0.5583  0.4603 -2.4218 -0.1881  0.4917
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]