Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-0.9644  0.2740 -0.4530  1.4593  0.8490  0.9326 -0.0306 -0.9126 -1.3109  0.4427
#>  0.1490 -0.2062  0.1147  0.6247  0.1252  1.8453  0.0805  2.1488 -0.2270  1.1918
#> -1.0430 -0.2595 -1.0706 -0.1893 -0.0195 -0.9070  0.3733  0.8495 -0.6710 -0.8513
#> -0.4174 -0.1639  0.2653 -0.9058  0.6341 -0.8195 -0.1863  0.2887 -0.9807  0.8492
#>  0.9282  0.1753 -0.8475 -0.3261 -1.4044 -1.0304  0.1169 -0.0274 -1.0040 -0.5732
#> -0.4338 -0.2061  0.2576  0.1084  0.0288 -0.4026 -0.3890  0.7186  0.9306 -0.2841
#> -0.7177  0.2161  0.6476 -0.4163  0.9690 -0.3569  0.0469 -0.5967  0.4329 -0.0398
#> -0.9512  0.2489 -0.2397 -1.8625 -2.2754  0.9314  0.0978  0.8082 -0.9741  0.7490
#> -0.3298  0.4839 -0.9406 -0.7803 -0.8235 -0.6578  1.0252  0.4606  0.9407 -1.3562
#> -0.9464  1.8291  2.3046 -0.3543  0.8567  0.9234 -1.3263  0.6830  0.9739 -1.3502
#> -1.7222 -0.6162  0.9303  0.4373  0.2778  1.4739 -0.6109  1.1165 -1.4623  0.4613
#> -0.3816 -1.4784 -1.3855 -0.6265 -1.2186  0.4282  0.8321  0.7988 -0.5596  0.0449
#>  1.3429  0.3188  1.0643  0.1368  1.2827 -0.6708  0.5263 -1.2821 -0.6134 -0.7399
#>  0.2449  1.7216 -1.2371 -0.2738 -0.7312  0.5273 -0.4843 -0.2365  1.3206  1.1022
#>  0.0945 -0.6168 -0.7211  1.3197  0.2884  2.3421  0.7718  0.9216 -1.1652 -1.3630
#>  0.2237  0.6777  1.6661 -0.3651 -0.3984 -1.0304  2.5098  0.5031  0.6383 -0.6501
#> -0.1724  0.5668  0.0401  0.1614 -1.6517 -0.2136  0.4057  1.0385  0.8213  0.2113
#>  0.2364  1.9207 -0.1202 -1.7840  0.6060 -1.6682  1.0681  1.1470 -1.1434 -0.7410
#> -1.4186  1.2247 -1.0351  0.5379 -0.8417  0.4194 -1.0788  0.6055 -0.0986 -0.8024
#> -0.1715 -0.4426  0.2222 -0.4863  0.6926 -0.4114 -0.1345  1.0477  0.0576  0.1967
#>  0.7129 -1.0078  1.1932 -0.2758 -0.7645 -1.5024 -1.8713 -0.1932  0.8513 -0.1626
#>  0.0484  2.0850 -1.3498  0.9475 -1.0049 -1.8266 -1.2421  0.3115  0.7115  1.6486
#> -0.0568 -0.6063 -0.5242  0.7916  1.2457 -0.0456  1.8352  0.7394 -0.0449  0.3768
#>  2.3700  0.0472  0.3035  1.3465  0.6808  0.3911  0.1090 -1.2202 -0.3324  1.2625
#> -2.2319 -2.3497 -0.4326 -1.9185  0.0709  0.1375  0.4241 -0.3329  0.2953 -0.9011
#>  0.2816  0.2200  2.6230 -0.4910  0.8892  0.8667 -0.4846  0.1587  0.6168  0.6883
#> -0.3223 -0.3723  0.1456  0.3220 -0.5395  1.4902 -1.8135  2.7462 -0.0056 -0.1582
#>  0.5650  0.4076  0.1769  0.5940 -0.7976  0.3002 -0.0187 -0.1367  0.6402  0.9437
#>  0.3473  0.8126  0.8715 -0.9136 -1.3066  0.5241  0.8796 -0.5798 -0.3592  1.0519
#> -0.1943  1.2244 -0.5180  1.1449  0.8428 -1.8226 -0.7356 -0.4748  0.3767 -1.3232
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]