Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3150  1.1126 -0.0293 -1.8887  0.3575 -0.1104  0.9168 -0.4910 -2.0614 -0.1819
#> -1.3241 -1.5535  0.3739 -0.3103  0.6003  0.2912  0.4791  0.4786  0.3377  0.6167
#>  0.2682  1.2085 -0.4763  0.0392 -0.7142  0.4646  1.0708 -0.2471  0.1371 -1.3233
#>  0.1029  0.5513  1.8328  0.6854  0.8637  0.4662  0.6144 -1.3120  0.7320  0.1005
#> -1.6664  2.8410 -1.4770  0.0153 -1.5578 -0.1455 -0.6827  0.3966  1.0701  0.7870
#>  2.2224  0.0652 -0.8217  1.3019 -0.7314  0.6526 -1.0356  1.1800  0.5405 -0.1939
#> -2.2055  0.1714  1.3492  0.3263 -0.6067 -0.1018 -1.2820 -0.2064 -0.1902  2.3771
#> -0.2617 -1.3217 -1.4613 -1.0204 -0.4474 -0.6590 -0.1412  1.4276  0.0491  0.1756
#> -1.8009  0.7213  1.6277 -1.1803 -0.2958 -1.1193 -1.2274  0.6893  0.5875 -1.7320
#>  1.5545  0.6961 -0.0733 -0.4250 -1.0578  1.0401  0.1386 -0.7211  0.6457 -0.8843
#>  0.0430 -0.2474 -0.5372  1.1296  0.3593  1.0911 -0.1369 -0.6860  0.1943 -1.1926
#> -0.0369 -0.7601 -0.9678  0.4146  0.0722  1.3735 -0.5456 -0.1807  1.9148  0.0110
#> -1.1750  1.3927 -2.2745 -1.2705  0.3270 -0.1346  0.4282  0.9025  0.6435  1.9298
#>  0.6280  1.3200  1.6727 -0.6872 -0.5586  0.5100 -0.6002 -0.8152 -0.8216  1.1586
#>  1.1226 -0.3479 -0.3412  0.1540  0.5624  2.0050  1.6304 -1.5482  0.2509  0.5057
#>  0.1463  0.0054 -0.0622  0.3908  0.1279  1.6800 -0.0112  0.2920 -0.0816  0.1340
#>  2.0712 -0.2670 -0.5934  0.1904 -1.2181  0.4419  1.0172  1.6760 -0.1881 -0.2747
#>  0.0676  1.1492  0.1435 -1.3830  0.1207 -0.7924  1.2008 -0.5633  1.5468 -0.5895
#>  0.5551 -0.7191  0.2235  0.7458  0.9379 -0.2171  0.2854  0.3533 -1.0532  0.5211
#> -0.0282 -1.1141 -1.2293 -1.2059  0.2263  0.0375 -0.1873 -0.3529  0.1904  1.9739
#> -0.4106 -1.0986  0.5715  0.1282 -0.9970 -0.9569 -0.4296  0.6925 -0.5251 -0.3527
#> -1.7924 -0.7144 -0.6166  2.0310 -1.0091 -1.9121  1.0211 -0.2583 -0.3374  0.6274
#>  0.5731  1.2428 -0.3180 -0.2386 -0.5966  1.3300 -0.8762  0.6070  0.7362  0.6155
#>  2.6065 -0.2005 -0.2216 -1.6032  2.6030  1.0776 -0.3641 -1.0161 -0.0657 -1.1579
#> -0.2808 -1.3894  1.0874  0.7927 -0.7059 -1.6137  0.3279 -0.8347  0.0304 -0.3634
#> -0.7441 -1.0222 -0.4886 -0.7941  1.7197  0.7119 -0.0659  0.5434  0.0273  1.1316
#>  0.3453 -1.0855 -0.4628 -1.4653  1.7238  0.4948 -0.7846  0.8323  0.3987  0.7553
#>  0.8934  1.2523  0.5653 -0.8159  0.0436 -0.8136  1.8424  0.0633  1.9820 -0.5423
#> -0.8900  0.0727 -0.1856 -0.1098 -1.6185 -0.2675 -1.9902 -2.0517 -1.0267 -0.7988
#> -1.7444 -0.8527  0.3424  0.8308  0.7959  0.3995 -0.3232 -0.6293 -0.7747  1.5505
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]