Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.2081 -1.2388 -0.3692 -0.2810 -1.1442 -0.6812 -0.7207  0.7423  0.3890 -0.3123
#>  0.9554  0.2274  0.0464 -1.2160 -1.2080 -0.9791  0.3590 -2.2287  0.2618 -0.5566
#>  0.4302  0.2629 -0.9153  0.4354  0.1243  0.0014 -0.7242  3.0715  1.4873 -0.2287
#>  1.1785 -1.6504  0.9056 -0.1177  0.2326  0.2705 -0.2712 -0.8177  0.8480  0.5388
#> -0.1726 -1.5647 -0.5706 -1.1071  0.9194  0.6132  1.5483 -0.0378 -0.1974 -0.1757
#> -0.6385 -0.8682  1.8558 -0.0579  1.8469 -0.5848 -0.0375  0.1150  0.7568  2.3715
#> -0.3288 -0.4429 -0.0458 -1.1874 -0.8812  0.7574  1.1478  1.3802  2.3817 -0.4644
#> -1.9219  1.2173  0.2555  1.6727  0.5571 -0.0824  0.8867  0.2658 -1.1271  0.6323
#>  1.0211  0.7621  1.0643 -0.1573 -0.3688 -1.1371  0.4760  0.3562  1.4343  0.3527
#>  1.3289 -0.3363  0.7413 -1.9396  0.2100  0.6029  1.1440  0.9438  0.3706 -0.0612
#>  0.3123  0.4028  0.0289  2.3512  0.2678  0.1122  0.6061 -0.5836  1.7983  0.4833
#> -0.3993 -0.4886 -0.8099  0.6658 -0.0181  0.1639 -0.4662  1.7730  1.8073  0.1786
#> -2.5239  0.0094 -0.1007  0.8594  0.4540  0.3087 -1.1495  0.7177  0.4345  0.2869
#> -0.1319 -1.0700  0.2889  1.9728 -2.1654 -0.8227 -0.1884 -1.0477  0.8728  0.4373
#>  0.1538  1.5011 -0.4775  0.6228 -1.4200  0.5050  1.2529 -0.1389 -1.8503  0.9055
#> -0.0316  0.1607 -2.1828  1.6599 -1.1054  0.0949 -0.0357  1.5870  0.2263  0.1660
#> -0.7215  0.9790  0.6446  2.4318  1.0821  0.0915 -1.2732  0.8224  0.1156  3.1063
#>  1.4475  0.1838 -0.0497 -0.6961  1.2401  1.8388 -0.9347 -1.4621 -0.0485 -0.0671
#> -0.5964  0.5896 -0.1412 -0.5488 -0.5448 -0.4081 -0.0389 -0.4702 -0.6063  0.0821
#> -1.7443 -0.8235  1.3158  0.5826  0.2755 -0.4617 -0.7319  1.3501  0.2447 -1.4233
#>  0.4170  0.2634  0.5371  0.7976  1.0899  0.2625  0.2931  0.6615 -1.9689  0.2854
#>  1.6519  0.5057 -1.0071  0.7987  0.1335  0.2334  1.5595 -0.7313 -0.3926  0.9833
#> -0.5758 -1.1232  0.1559 -0.2349  0.4413 -1.3987  0.7871  0.7075 -1.2259  0.7116
#> -0.1631 -0.1287  0.1541  0.7591 -1.8113 -0.2026  1.0474  0.0029  0.8781  2.9457
#>  0.2865  0.5908  1.3901  0.6248  0.4499 -0.7894 -0.4529  1.3642 -0.6161  0.6754
#> -0.2662 -0.3387  1.8827 -1.6493  0.1531  0.9907 -0.5355 -0.6519 -0.0177 -0.0421
#>  1.1206 -1.1640 -1.0406 -0.9061  1.6627 -0.3456 -0.7903 -0.5080  1.2817 -0.3939
#> -0.5792 -0.3952  1.3580 -0.0257  0.7260  2.2829  0.3632  0.8225  0.3401  0.2157
#> -0.4066  0.5881  0.4035  0.8700  0.9212  0.3522  0.8435  1.2747 -0.4062  0.0949
#>  0.1737 -1.0004  0.0854 -0.0437 -1.7020  1.5102 -1.3633  0.3995 -0.4233 -0.1614
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]