Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7255  0.5674 -1.9588 -2.2001  0.3433 -2.0595 -0.6125  0.4461 -0.1759  1.5046
#> -1.0679  0.2093 -0.4542  0.7235 -0.8087  1.5666  0.5329 -2.4741 -0.2915  0.7113
#> -0.9947 -0.2839  0.6779 -2.1238  0.1170  0.3864 -0.3429  0.1735 -1.0263  1.0916
#>  1.0272 -0.3691 -0.0024  0.8412 -0.0034  0.0412 -0.4070  0.8485  0.2808  0.0379
#>  0.4543 -0.9269  0.5789  0.3829  1.8657  0.2578 -0.6467 -0.2380 -0.3442 -0.3245
#> -0.1171  1.5339 -0.3661 -0.0279 -1.0966 -1.1020 -0.9627  1.3354  0.1944  1.5851
#> -1.0636 -1.4793 -0.3620  0.3484  0.7581 -0.7424 -1.1663  1.3757 -0.6015  0.6850
#>  1.3001  0.9135 -0.1669  1.0226 -1.0255  0.2304  0.3848 -1.2220  1.0660 -0.1391
#>  0.2303  0.2090  0.5745 -0.1453  0.8973  0.9941 -0.9287 -0.1396 -0.7759 -0.1780
#>  0.1617  1.1546  1.1653  1.2536  0.1215  0.9692  0.4111  0.8015 -0.8062 -0.5138
#>  0.8664 -2.2262  0.5185  0.4682  0.6870  0.3686  0.9160  0.4800 -0.1689  1.3330
#>  0.1024 -1.0737  1.3005  1.0660 -0.0474  0.7495 -2.0304 -0.6188  0.7161  0.1383
#> -0.8776 -0.0905  1.1585  0.1237  0.1008 -0.0175  0.6864  1.5129  0.4352  0.1663
#> -0.6141 -2.2606  1.7681  0.7573 -0.6160  0.3019  0.4645 -1.4387  0.8300 -1.3829
#>  1.7569  1.2960  0.8463  0.4368 -1.0621  1.2440 -0.9770  0.0806  0.5809  0.0166
#> -1.2608 -1.4460  0.2329  0.1447 -0.1024 -0.2783  0.5567 -1.0051  3.1195 -0.7193
#>  1.2938 -0.5378 -0.6987 -1.6153 -1.0524 -0.1229 -0.3999  0.5930  2.0718 -1.4884
#>  0.3328 -0.5898 -0.3931 -0.3759  0.1808 -0.7923 -0.5170 -2.2846  0.6754 -0.9715
#> -0.0800 -0.1150 -0.0146  0.4072 -0.2781 -0.3415  1.1795 -0.6471  0.8994  1.3045
#>  1.8510  1.3744 -0.3242 -1.6394 -0.6489 -0.1657  1.3181 -0.0089 -0.2250 -1.8010
#> -0.3325 -0.2073 -0.5520  0.2486 -1.1350  1.3677 -1.6427  1.6643  0.8786  0.2897
#>  0.0695  0.6530  0.9735  1.1907  0.0193  0.7689  0.3223  1.2653 -0.8809  0.7421
#> -1.7331  0.2347 -0.3392  1.7112  2.0109  0.5558  1.1813  1.3735  0.5318 -1.2895
#> -0.6800 -0.9712  1.5020  0.7171  1.5082  0.9254  0.3699 -0.2717 -0.2642 -0.4296
#> -0.5248  0.6756 -1.4471  2.0288  0.1099  0.3717  2.8610 -0.2807 -1.0066 -0.7300
#>  0.2518 -0.5436  2.2542 -1.0685 -1.0655  0.0806 -0.2520  1.3488 -0.0891 -0.3780
#>  0.3880 -0.8907 -0.0890 -0.8900 -0.8824  0.0144 -0.9509 -0.3333 -0.1506 -0.6154
#>  0.1502  0.4390  0.3134  0.3108 -0.5720 -0.0520  1.1053  0.6866 -0.0662 -1.8749
#> -0.6238  0.0342 -0.7166 -1.1620  0.2010  1.3162  0.9391 -1.5733 -1.4300  1.4163
#>  0.3994 -0.8007 -2.1268  1.8969 -0.0738 -0.3783  1.4178  1.4154 -0.4145 -0.0372
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]