Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.1015  0.7097  0.5150  1.9079  0.3963  1.1650  0.1297  0.4841 -0.6537 -1.0504
#>  1.0518 -0.0500 -1.1850 -0.0658  0.4525 -0.8628  0.0206 -0.6715  0.6054  0.4980
#>  0.3629 -0.2711 -0.4230  0.3462 -0.2618 -0.7584 -0.6359  1.1159  1.2498  0.1426
#>  0.4982 -0.6625 -0.6261  0.6774 -0.7440 -1.7563 -0.0996 -1.2265  1.6801  1.3938
#> -0.1422 -0.7648 -1.2834 -0.9274  0.4678  0.3962 -1.3818  2.1012 -1.6696 -0.5034
#>  3.1335  0.1996 -2.0323  2.1150 -0.5630  2.1974 -0.0546  0.3395 -0.0097 -1.2267
#>  0.2194  0.2028 -0.6478  0.0980  0.2199  0.2070  1.5819  0.1953 -0.1081 -0.9809
#> -0.9320  1.5231 -0.4954  1.2382  0.2673 -0.2882  0.5749  0.6711 -0.8834  0.6196
#> -0.6141  0.7820  0.6768 -1.3209 -0.3975  0.0040 -1.2860  1.9482 -0.0686  1.1730
#> -1.6043 -0.2293  1.2690 -1.5849  1.0259  0.2617  1.2617  0.4453  0.0957 -0.0765
#> -1.1753  0.9852  2.0215 -2.1855 -0.5624  0.3355 -1.8006 -0.2845 -0.1129 -0.6466
#>  0.3651  0.7058 -1.2109  0.0011 -2.1433 -0.6151  0.3742  0.9679 -0.9004  1.6725
#>  1.0647  1.0762 -2.2534 -0.2475  0.8045 -0.0596 -0.7730  0.5547  0.1348  0.4044
#>  1.0623 -1.1155 -0.0666  0.1352 -0.3053 -0.7032  1.7282  0.0647 -1.6354 -0.4833
#>  1.2542 -1.0177  0.1215 -0.9165 -0.8646  1.0032 -1.8859  0.3465  1.6223 -1.2074
#> -0.8071  0.1878 -0.1394 -0.3981 -1.3620 -1.1771 -0.4578 -0.0557 -0.8512  0.4960
#>  0.5039  0.9606 -1.9320 -0.0956  0.4994  0.0723 -0.1186  0.2496 -0.5693  0.2339
#> -1.3946 -0.5504 -0.0321 -0.7941 -1.4054  0.3957 -1.3178  0.2307 -1.3567  0.7153
#>  0.1909  0.3319  0.6631 -0.0545  0.9553 -0.8330  0.5847 -0.6222  1.2243 -1.3043
#> -0.9727 -1.0357 -1.2140 -1.6395 -2.0273  1.1510  0.0837  1.9412 -0.8716 -0.0639
#> -0.5444  0.1872 -0.0386  0.7809 -0.0985  0.6584 -1.5865 -0.9716 -0.0774  0.9525
#>  0.0648 -1.1354  1.2389 -0.2093  0.1814 -1.5235 -0.4200  0.0705 -0.4636 -1.0960
#> -1.2200  1.8605  1.0736 -0.5924 -0.0341  0.6797  1.3380  0.4381 -0.7883 -2.4753
#> -0.3208 -0.1740  1.7337 -0.5718 -0.5112 -0.3898 -0.4082 -1.1541 -0.0979  0.1747
#>  0.4890  0.2132  0.2638 -0.9494  0.5147  1.9392 -1.1982 -0.0287 -0.6402  0.5957
#> -1.1015 -0.3996 -0.5641  0.5326 -0.7426  0.8540  0.8373  0.2824  0.3029  1.2779
#>  0.4256  1.5561 -0.1470 -0.9909 -0.6087 -0.0670 -0.3075  0.7078 -1.5082  1.0747
#>  1.8222  0.5456  0.2187  0.1160 -0.3680  0.3340  0.8530 -0.2242  1.0797 -1.4286
#> -1.5686 -1.1406  1.1739 -0.3837 -0.9908 -2.5793 -1.1178 -0.3860  0.5358 -1.1449
#> -1.1757 -0.1618  1.2660 -1.0796  0.6840 -0.1946 -0.3538 -0.4729  0.7671  0.0865
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]