Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.9261  0.9418 -0.7778  1.2233  0.7476 -1.1140 -0.7096  0.0032  1.0958  0.3457
#>  1.4508 -0.8524 -0.1514 -0.3074 -1.1394 -2.1068  0.8387  1.2977  0.0616  1.1716
#> -1.9935  0.9931 -0.5216  0.1458 -0.9152 -0.3737  0.6524  1.8460  0.6564 -0.3106
#>  0.4749  0.0072  1.0387  0.2198 -1.9050 -0.9680  1.6809  0.1686 -0.6777  0.8706
#>  2.6400 -0.1626  0.1998 -1.9508  0.0261 -0.9145  1.6028 -1.4151 -0.0260 -0.1157
#> -0.4798 -1.3444  0.8803  1.4800 -0.4710  1.9553 -0.3206 -1.2624  2.5098  0.4436
#> -0.9610  1.1119  1.0995  1.2528  0.0277 -0.0230 -0.6052 -1.6956 -1.0007  0.5122
#>  0.0415 -0.6612  0.2294 -1.1595 -0.2518  1.0827 -1.1921  1.0308 -0.7013 -0.8091
#>  1.3133  0.4160  0.6887  0.6210  1.3611  1.2013  0.6739  0.6645 -0.2138 -0.0306
#>  0.1505 -0.3234 -0.5667 -0.8266  0.6032 -0.9945  0.9409 -0.5360 -1.2146 -2.6901
#>  1.2875  2.3264 -1.2686  0.9034  0.6943 -0.0550 -0.1540 -0.1610  0.5305  0.2888
#>  0.6114  0.4533  1.5405 -0.7132 -0.8143 -0.8858 -0.7930 -2.6947  2.4976 -0.3237
#>  0.8098  0.9457  0.6687  0.8437 -0.6024  1.4155  0.0988  0.7178  0.0719 -1.5385
#> -1.2818  0.3341 -1.3622  1.5548  0.3828 -0.3009  0.0574 -0.1607 -0.2189 -1.7074
#>  1.9249  0.6952 -0.6009 -1.6085  0.3634 -1.3780  2.0431  0.8597 -0.7062 -0.7164
#>  1.8453 -0.5021  0.0116 -1.0247  0.0943 -0.7737  0.6847  0.3627  2.2620  0.3475
#>  0.5583 -0.4883 -1.6718 -0.8798  0.5841 -1.3192 -0.3149 -0.7963  0.1672  0.0165
#> -1.4700 -0.3869  0.3963 -0.3537 -1.9272 -0.0024 -0.3405 -0.1196  0.7867  3.2541
#>  0.4477 -1.2556 -0.1436  1.5514 -1.0465  0.1043 -0.0108  0.8989 -0.9046 -0.3271
#> -0.9114 -1.2265  0.9148  1.4380 -0.2810  0.3265  0.5354 -0.6412  0.4129  0.7983
#> -0.1150  1.9264 -2.5068 -0.1522  1.3075  0.7677  2.2331  0.0077 -0.1134 -1.2793
#> -1.5644 -0.4271 -0.1402 -0.0966 -0.7031 -0.5964  1.1162  0.4938 -0.3712  1.8761
#>  1.3501  0.4068  0.6611 -0.4085  1.2081 -1.3370  0.9821 -0.7803  0.9860 -0.7644
#>  1.8496  1.5315 -0.5749 -1.1618  1.2279 -0.7506 -1.3306 -0.8432  0.5307  0.2364
#>  0.6103  0.8462 -1.3727 -0.2869  0.0807  0.3106 -0.4122 -0.5511 -0.9832  3.3790
#> -0.0166 -1.2455 -1.5213 -0.8329 -0.3635 -0.4961  0.4912 -0.1843  1.5673  0.5836
#>  0.7490 -0.4302 -0.2268  0.7308  0.1404 -0.2474 -0.9911 -0.4256  0.5478  1.2142
#> -0.9427 -0.2197 -0.5307  0.0568 -1.1071  0.5264  0.3913  0.2609 -1.0265  0.4758
#>  0.6234 -0.0738 -0.6179  0.0884  0.5784  0.4244 -1.8581  0.6324  0.2013 -1.4051
#>  0.1144  0.0141 -3.0155  1.8224 -2.4322  1.3758  0.8299 -1.2332  0.5685 -0.6270
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]