Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.8024 -0.7317  0.4978  1.3937 -0.0921 -0.0523 -1.9347  0.0231 -0.6795 -0.2087
#> -1.7025 -0.8328 -0.6750 -0.5064 -1.7727 -0.1646 -0.1597  0.1703 -0.6181 -1.9692
#> -0.0819  2.5484  0.4617  0.2579 -2.8889 -0.1514  0.3622 -0.7758  0.6303 -0.9392
#> -0.0547 -0.4780  0.1759  0.8688 -1.6427  0.3991  0.7996  1.2872 -0.8394  1.4218
#>  0.0107 -0.1774  1.5139 -0.8151  0.5054  0.1397 -2.3107 -0.2646 -0.8029  1.1065
#>  0.0041 -0.0564 -1.6040  0.4779 -1.6337 -0.8372  0.0262  0.7832  0.3668  0.3416
#>  1.5601 -0.8907  0.6612 -1.7718  2.5405 -0.2843 -0.4901 -0.8556 -1.8308  0.5488
#>  0.0141  0.3080 -0.9695  0.0091 -1.0617  1.2131  0.2604  0.4306 -0.2082  1.0026
#>  0.6578  0.7249  0.6743 -1.2540 -1.1217  0.1685 -1.3356  1.4835  1.3899  0.9818
#>  1.3000  0.0894  1.9533 -1.3134 -0.3265  1.8296 -0.4756 -2.7403 -0.1670 -0.2217
#> -1.1765 -0.8157  0.3299  0.4823  0.1484  0.5174  0.3578 -0.6888 -1.5955 -0.7140
#> -0.4479  1.3062 -0.3480  0.4310 -0.4028 -2.2206  0.4493  0.3969 -0.7924  1.2675
#>  0.0432  0.0076 -1.9295 -0.7706  0.4904 -0.5190  0.0766  1.7864 -0.0682 -0.0769
#> -1.5958  0.8349  0.2191  0.8593 -0.5417  0.7741  0.1767  0.4168  0.1641 -0.2807
#>  0.5010  1.0239 -0.5809 -0.9421 -0.0356 -1.7316  0.5134  0.0832 -0.5861  1.2667
#> -0.5946 -0.8740 -0.7528 -0.4375 -1.3711  0.5207 -0.7489  1.5951 -0.3965  0.5755
#> -0.6460 -0.0204 -0.3402 -0.4195  0.1003  0.2330  2.6882  0.1583 -0.3522  0.3836
#>  2.0999 -1.5478  0.1421 -0.9419 -2.0544  0.4318  0.7141  0.0093  1.0695 -0.8079
#>  1.0128  1.3111  1.4588  1.8842  0.8013 -0.9287 -2.0523  0.1684  0.0095  0.3045
#>  0.5912  0.2561 -1.0270 -0.8324 -1.1616  0.4904 -0.4202 -0.9161 -0.8267 -1.3195
#>  0.5510  0.5217 -1.0639 -2.2729 -0.7851  1.0556 -1.3054  1.7103  1.2033  2.4333
#> -1.5995  0.0285  0.3102 -1.6101  1.0147 -0.7734  1.2279  0.5967  0.9507 -0.2208
#>  0.7198  0.6204 -0.6281  1.3101 -0.9580  0.0177 -0.1845  0.1396  0.0949 -1.1699
#>  1.0812 -1.0853 -0.4804 -1.9239 -0.5047  0.3429 -0.5791 -0.8595 -0.1190 -0.3236
#> -0.0483  0.6408  0.4575  0.0460 -1.7740 -1.7590  0.7074 -0.9135  0.0788  2.1875
#> -0.4634 -0.6389  1.3440 -0.4923 -0.7067 -1.3406 -0.2863  0.5032 -1.6412  0.2790
#>  0.2083 -0.4312 -0.8936  0.1721  0.9697 -0.4812 -0.5157  0.6378  0.5468 -0.1255
#> -2.6140  0.4905 -1.8341 -0.8477 -0.6288 -1.5263 -0.7875  0.4241 -1.6891  0.6970
#>  0.0826  0.1351 -1.3822 -0.5184  1.5562 -2.4186  0.8563 -0.5489  0.5505 -1.5283
#> -0.1108  1.0991  0.3744 -0.2980  0.2565 -0.1999  0.8797  1.9200  0.8952  0.2992
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]