Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2394  0.7526  0.2326  0.3667 -0.3447  0.4717 -0.3187 -1.8909  0.7091 -0.1670
#>  0.0482  0.8080 -0.7065 -0.2927 -1.5460 -0.7640 -2.1165 -0.2170 -0.7271  1.7099
#>  0.4382  0.2405 -0.4564 -0.8742 -1.2384  0.0008 -0.5322 -0.9267 -0.5769  0.1801
#>  1.8094 -1.3734 -0.2132 -0.5010 -1.2386 -0.4799 -0.2550 -0.8569  1.2539  0.4527
#> -1.9001  0.0772 -0.3601  1.4497  0.7648 -1.5969 -0.7821  0.9289  0.0475 -0.5949
#> -1.5539 -0.1442  1.3423  0.4099 -0.8735  0.8511 -1.1524  0.3175  0.1501  0.2601
#>  1.1435  0.0066  0.3543 -1.6489  1.9844  1.2107  0.1256  0.2553  1.1367  0.4046
#>  0.0181  0.0847  0.5669  0.7953 -0.1546 -1.5408 -0.0324  0.9451  0.7123  0.1397
#> -0.0760  1.0732  0.1066 -0.8163 -0.9877 -0.5829  1.6171  0.4151 -1.0586 -0.7713
#> -0.9101 -2.3319  1.0384 -1.5131  1.2588  0.7043  0.6570 -1.1143 -0.8114  0.9647
#> -1.1326 -0.7346 -0.5852 -1.3603  0.5897 -0.0976  1.4430 -0.2954  0.8627 -0.7187
#> -1.1068  2.0039  0.4673 -1.2975  0.1805  0.9112 -0.8653 -0.5212  1.4594 -0.4028
#>  1.3940 -0.5852 -0.4186  0.7044 -1.1765  0.1196  0.8585 -0.3175 -0.5905 -0.9471
#> -0.8477 -1.0604  1.5680 -0.7008  0.1781  0.0847 -0.3257  0.6126  0.4932 -1.4677
#> -0.4956  0.8537  0.6873 -0.5940 -0.3126  0.0975  0.2623  0.5971  0.2019  0.4166
#>  0.8307 -0.9470  0.7725  0.5585 -0.3065  0.5132  0.2152 -0.3903  0.4105  0.5113
#> -0.1745  0.1810 -1.8835 -1.2769  0.0180  0.8614 -1.4547 -0.4307 -1.5404 -1.4544
#>  0.4415  0.3050 -0.8433  1.3566 -2.8365 -1.6793  1.0721  1.7066 -0.7452 -1.1157
#>  0.0652 -0.2886 -0.2196 -0.6655  0.5353  1.5799  0.3794  0.3166  0.9173  0.6483
#>  0.5059  0.9056 -0.1759  2.6114  2.2896 -0.3638 -0.1609 -1.2563  1.6194 -1.4232
#>  0.4498 -0.3232  0.0218  1.7377  1.2332  0.9277  0.5636 -0.5969  0.8547  1.7300
#>  1.0271 -0.7969  0.8827  2.4251  0.0683 -0.8179 -3.1973 -0.2346 -0.1015  1.3891
#> -0.3069 -0.4415  0.5533  0.1260  0.4511 -0.4213  1.8306 -0.0624 -0.2472 -0.0928
#>  0.6064 -0.1537 -0.1296  0.6448  0.2442 -0.3082 -1.0421 -0.7551 -0.3952  1.5525
#>  1.5060  0.4031  0.1421 -0.0989 -0.9369  0.4491  0.9992 -0.4707 -1.2291 -1.0397
#> -0.6133  0.0311  1.2283 -0.4497  0.1781  0.8160  0.5965 -1.2330  0.9968  0.2511
#> -0.8008 -0.8617  0.9809  1.2056 -0.8079  0.4900  0.3194  0.2364  0.4146  0.4394
#> -0.0620  0.3618 -0.0332 -0.2596  0.7945  1.1740  0.1474 -0.0693 -0.1262 -0.9485
#> -0.2147  0.6029  0.2441  0.4661  0.0300 -0.1595 -0.1459 -0.8150 -0.1441  0.9812
#> -0.4389  1.4546  0.2312 -0.6937  0.5783  0.3690  0.1277 -0.2679 -1.0256 -0.7694
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]