Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.8576 -0.3445 -1.6549  0.5818  0.7884  0.1649  0.4753  1.5626  0.1260  0.1123
#>  0.8853  1.9478 -2.8269 -0.5027 -2.1136  1.6489 -1.1455  1.2394  0.2601  0.5048
#> -0.4583 -0.0486  0.1645  0.3839  0.9004 -0.5697  0.7650 -2.2509 -0.8675 -1.8368
#>  0.4690 -0.1317 -0.8884  0.4409  2.4203  0.8150 -1.7286  0.0185 -1.1919 -0.0259
#>  0.5805  0.6779  0.7176  1.2219 -0.7992 -1.2593  0.2926 -0.5525  1.2016  0.0658
#>  0.3753  1.3193 -0.5954  2.9050 -0.3593 -1.1683 -0.5427 -0.7265  0.5041 -1.5319
#> -0.8349 -1.5035 -0.3676  1.0358 -0.8660  0.5907 -0.7553 -1.4211  0.4167 -0.9291
#>  1.1094 -0.4955 -0.3355  1.8130 -0.3033  1.8345  0.3270  1.4217  0.3699 -0.8047
#>  1.8153 -1.1953 -1.4130 -1.8989  0.1375 -1.0711  0.4487  0.2338 -0.3032 -1.8887
#>  0.4473 -0.2179  0.1414 -0.0372  0.4060 -0.1261 -0.6376  1.9056 -0.1718  0.2784
#> -1.2271 -0.0816  0.0847 -1.4366  0.1337 -0.9225  0.5300  0.0176  0.9173  0.1213
#> -1.1045 -0.3521  2.2704  0.2499  0.3848 -0.3816  1.2307 -1.4470 -0.4337 -0.9328
#>  1.2745 -0.6117 -1.8735 -0.1440  1.6136 -0.6767 -1.9927  1.0944  0.3509  0.2955
#>  0.2602  1.4166 -1.0219  0.7007 -0.2331  1.7301  0.6326  0.7083  0.3014 -0.3680
#>  1.7293  1.9111 -0.2376  0.7873 -1.5579  2.3484 -0.1392 -0.6816 -0.0766  0.6550
#>  0.3667 -0.1494 -0.6067 -0.9852  1.2937 -0.2592  0.9406  0.5600 -1.4427 -1.8551
#>  1.4315 -0.3539  0.3628  0.7353 -1.7778  1.2308  0.5521  0.1702 -0.3263  0.5271
#> -0.2102  1.5453  0.3498  0.2899  0.6229  1.4839  0.1966  0.6682  0.2649  0.8476
#> -1.3524  0.8174  0.2369  0.4451  0.1070 -0.5323  1.4189 -0.1969  1.1107 -0.7885
#>  0.5337  0.5040  0.6727  0.8376  2.5929  0.6275 -0.3111  0.3829  1.3444  0.1599
#>  0.8232 -1.6955  1.3072 -0.1496  1.1104 -0.4422  0.1829  0.0472 -0.1317 -1.7097
#>  0.1779  0.0735 -0.2528  0.9983  0.6431 -1.0973 -0.8057 -0.6210 -0.5561  1.3519
#>  1.3613 -0.1872 -1.6190 -1.5717 -0.0390 -0.0799  0.8352 -0.9623  0.2721  0.1379
#>  0.0419 -0.2992 -1.1427  0.9799  0.1088 -1.0717 -1.3101 -2.8008  0.1535  1.4313
#> -0.2385  0.2543  2.3031 -0.2193  0.2301 -1.2736  0.5997 -0.2973 -0.7919  0.0776
#>  1.2330 -0.4694  1.1323 -0.2344  0.1287 -0.2679  0.1338  0.4414 -1.2678  1.2206
#> -0.6806  0.2628 -0.2276 -1.2071  0.1771 -0.3195 -0.9286 -0.6838  1.8832  2.0065
#> -0.2668 -1.1168 -1.0119 -0.4966  0.8321  0.2117 -1.3594 -0.3587  1.2503  0.2344
#>  0.6887  0.6249 -0.9533  0.4495  0.2611 -0.2930 -0.5011 -1.7623 -0.1734  0.9399
#>  0.4961 -1.0710 -0.5876 -1.3578 -1.3919 -0.8088  0.9865  0.7205 -1.0635 -0.2879
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]