Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6069 -2.4275 -1.3992  0.8521 -0.3858  1.0238 -0.9715  0.5830 -0.0963  1.0923
#> -1.8037 -2.5131 -0.2842 -0.2400 -0.0204 -0.5489  2.1293 -2.7470  0.1472 -0.1866
#> -2.2566 -0.0400  1.2418  0.9707  0.2315 -0.9929 -0.3285 -1.1366 -0.5214 -0.2062
#>  0.2773  0.4117 -1.3119  1.2059  0.5434 -1.0274 -1.3323 -0.3289 -0.1906 -0.3772
#> -0.6394  1.1143  1.5466  0.8171  0.3659  0.3869 -1.0849  0.3327  0.6738  0.5152
#> -0.8083  1.5755 -2.1427 -0.1641 -0.0991 -0.0243 -0.3704  0.2774 -0.2978 -2.4672
#> -1.0830 -2.0681  0.2745 -1.7779  0.7938  0.8498  0.6512  0.2258  0.6906 -1.1262
#>  0.1185  0.0904  0.9939 -1.2590 -0.4790 -0.3694  1.1316  0.3528  1.4597  1.0126
#>  1.2643 -0.0104 -1.1152  0.0010  1.2979 -0.2335 -0.7909  0.5500 -0.6157 -0.7037
#> -0.8197 -1.2421 -0.8537 -2.2402 -1.1850  1.2200 -0.0710  0.9816  0.0105  0.2912
#> -0.7721 -2.1041 -1.3376 -1.7983  2.3732  1.2260  0.5395 -0.6348  0.7037  0.7998
#> -0.0539  0.4824 -1.2551 -1.0681  0.9319 -0.0407 -0.3514  1.4988 -0.9891 -0.5289
#> -0.5132  0.4088 -0.1557  0.2339  0.8199  0.3555 -0.2555  0.5971  0.0468  1.0915
#> -0.5985  0.5276 -0.8725 -3.2618 -0.6484 -0.2208 -0.4747 -0.7315  0.7265 -0.8420
#> -1.3179 -0.2555  0.4684 -0.7771  1.7786  0.7261 -0.6200  0.5748  1.9644 -0.4337
#>  0.1988  0.1467  0.2206 -1.4495  1.2272  0.1630 -0.5341  0.0260 -0.3829  1.3280
#>  1.3680  1.0332 -1.1620 -0.0127 -1.0797  0.6563  0.4661  2.6786 -0.9325 -0.8821
#>  1.0408 -0.0016 -1.8150 -1.6612  0.2391  0.8189  0.8911 -0.6445 -0.8748 -0.0502
#>  0.0021 -0.6094  1.4809  0.5708 -0.1572 -1.7370  1.0455 -0.5582 -1.3724  1.7043
#> -1.1273 -0.8495 -0.4839 -0.1063  0.3365  0.4505  0.5936  0.4891  1.0807 -2.1122
#> -0.9469 -0.7991  1.2373  2.9354  0.1893  0.7385 -0.4514  0.7722  0.6687 -0.5433
#>  1.1359  1.1222  1.2331 -2.3933 -0.3234  1.5387 -0.6643  1.5983  1.1301  0.2268
#>  0.3987 -0.6858 -1.0594 -0.5784 -0.5806  1.4915 -2.4708  0.7126 -0.9440  0.0070
#> -0.2353  0.5885  0.3164  0.5740 -0.6350  0.2513 -0.8294  0.0914  1.9731 -1.7522
#> -0.2699  1.2086 -0.5492  1.0868 -0.3500 -0.1949  0.0191 -1.1240 -1.1418  1.4128
#> -0.2845  0.8168 -1.3844 -0.4617  1.0185  0.6746 -0.4667 -1.1741  0.1102 -0.4092
#>  0.4074  0.2555 -0.3664 -2.0780 -1.6050 -1.0317 -0.1623  0.1343  0.7066 -2.0259
#> -0.9460 -0.4092 -1.5586  0.4690  0.6270 -1.0172 -0.6085 -0.1201  0.5084  1.2628
#>  0.6348 -1.4996  0.4127 -0.5767  0.3381 -0.2992  0.8794 -1.8930  0.5295 -1.4683
#> -1.3587 -2.2363  1.4297 -0.2047 -0.8460  1.5056 -0.6968 -0.0857 -1.6934 -1.0093
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]