Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.9178 -1.6675 -0.2624  0.4261  0.6908 -0.0500 -0.2154  0.6296  0.4183 -0.1439
#>  0.8490 -2.0586 -0.4890  0.8020 -0.4133  0.5236  0.1528  1.2948 -0.9466 -0.5462
#> -1.1863  0.1001 -0.0096  1.3305 -0.8081  0.5912  0.4039  0.5829 -0.5587 -0.3671
#>  0.0686  0.2341 -0.1356  0.0984  0.5946  1.3504  1.2616 -1.2791 -1.5939  0.4556
#>  1.3230  1.5903  0.4649  0.5500  1.3761  0.9472 -0.1673  0.5920 -0.3537 -1.2156
#>  1.2531 -0.1604  1.7761  0.2896 -0.8404 -0.1748 -0.1397 -0.4530  1.3067 -1.4998
#> -0.1458  1.0912 -0.0568 -0.2552 -0.9540  0.0815 -0.2839 -0.8746  0.1362 -0.9845
#> -0.0589 -0.1258  1.8686 -0.8643  1.0980  0.4601  1.1919 -0.8905  0.3733 -1.7232
#>  1.4959 -0.3847  0.8758 -0.8343 -0.3542  0.5565  0.9270  0.4582  0.0529 -1.6927
#> -0.3714 -0.1896 -0.2291  0.4058 -0.6336  0.9528  1.2159 -0.6322  0.1961 -0.5223
#>  0.7947 -0.0871  1.8865  0.3815  1.4515  0.5377  0.5684  0.7694 -0.9905  1.9295
#>  0.9622  0.7320  0.4062 -1.2008 -1.0601 -0.5397 -0.4740  2.0267 -1.0028  0.1733
#>  0.4022  1.6707  0.1412 -0.1771 -0.4321 -1.8470  0.9528 -0.5378  0.1186  0.3146
#>  0.5180  0.7190  1.0150  0.2814  0.4391  0.9243 -0.7749 -1.3005 -0.0051  0.4544
#>  1.2729 -1.6603  0.5952 -1.0343 -0.9043 -0.5702 -0.2928 -0.1904  1.1185 -0.4507
#> -0.9650  1.0316 -1.4370  1.1346  1.3083  0.4176 -0.6156  0.0610  0.4051  0.3922
#>  0.6304  0.6018 -1.4673  1.5654 -0.4208 -0.9589 -1.4932  0.5358 -0.5082  0.0141
#> -0.1010 -0.6301  0.0124  0.6430  1.1722 -0.5136 -0.0029  0.8376  2.4653  0.9260
#>  0.3892  0.5418 -0.9269 -1.4043 -0.5503 -0.2028 -1.6691  0.6644 -0.2257 -1.0286
#> -1.2096 -0.3515 -1.8725 -0.5794  0.8223 -1.4928 -0.6237  0.8505  0.8421  1.9215
#>  1.9040 -0.8976 -1.4254 -1.3086  0.4668 -0.2336  0.0449 -1.0206 -0.4277 -0.5670
#>  0.8141 -1.3351  1.9282  1.3965  0.2316  0.1889  0.0762 -0.9386  0.1571  0.8476
#> -0.7556  0.2560  0.2756  1.1088 -0.9242  0.3863 -0.3114  0.5091  1.0899  0.5352
#> -0.9642 -0.5220  0.3652  0.8306 -2.2661  0.2274 -1.7612 -0.1545  0.3668 -0.6889
#>  0.2602 -2.1312  0.8257 -0.6231 -0.2394  0.2558  0.2164  1.4271  0.6501 -0.8449
#>  2.5338 -0.3063  0.0007  2.3488  1.5413  0.5368 -1.0297  0.7644 -0.2756  1.0421
#>  0.6627  1.0174  0.7555  1.0310 -0.3986  0.5735 -0.2958 -0.7105 -0.0529 -1.6314
#> -0.7570 -0.1666  0.0590 -0.1996 -0.2353  0.0837 -0.7269 -0.2372 -0.2111 -0.4009
#>  0.5645 -1.1000  0.3188  1.0711 -0.2039 -1.5551 -1.4456  0.7414  0.4435 -1.0681
#>  0.9112 -0.0075  0.2953  1.3041  0.5553  0.4002  1.1822 -0.5807  0.0145 -0.8870
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]