Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.9087  0.8906  0.2985 -0.4234  0.1628 -1.3172 -0.5408  0.5644  0.2150 -1.9326
#> -0.6616 -0.2766 -0.7988 -0.6522 -0.4346 -0.6186 -1.0075  0.0097 -2.0464 -0.5984
#> -1.5418 -0.6028  0.0722 -0.0315 -1.5120  0.0570 -0.3171  0.4464 -1.7025  2.1941
#> -0.3544  0.5222  1.7070 -2.2790 -1.1234  0.0763 -2.0158 -1.8245  1.1138 -0.4488
#>  0.1376  0.6180 -0.4316 -1.3437  0.6551  1.0499  0.9375 -0.6379 -0.1489 -2.3979
#> -0.6325  0.2602 -0.1654  1.0271 -0.2096 -0.3025 -0.3698 -0.6444  0.7659 -0.2137
#>  1.3683 -0.3563 -1.5087 -0.1181 -0.5553  0.9654 -1.5346 -0.3490 -0.1281 -0.0874
#>  1.6245 -0.4013  0.0530  0.1400  0.1749  0.7677  0.2764  1.2725 -0.4036 -0.6048
#> -0.5758  0.0511 -0.2637  0.2535 -2.9316  0.3896  0.3759  0.7285  1.4243  0.2227
#> -0.5467  0.0691  1.0938  0.5610 -1.2012  0.6247 -0.6892 -0.4695  0.9691 -0.1232
#>  1.4275 -2.1960 -0.6010  0.1306  1.2646  2.0836  1.5017  0.7477 -0.0136  1.3054
#>  0.3149  1.5517 -1.8320  0.0402  1.2466 -0.4622 -0.5759  0.0157  0.3580 -0.1192
#> -0.7769  0.1233 -0.3355  0.8225  0.9002 -0.0784 -0.5536  0.9825  0.7355 -0.9730
#> -0.0201  0.8647  1.9687 -0.7958  2.6018  1.4371  0.1673 -1.1849  0.0357 -0.9946
#>  0.0275  2.8176  1.3039 -0.6712  1.0662 -1.4585 -0.4366 -1.1998 -0.0656  0.6900
#> -0.3045  0.1062  0.5810 -0.3340 -0.8445 -0.2100  0.2927  0.7288 -0.4883 -0.4698
#>  2.5606 -1.3122 -0.2836  0.5343 -0.0104 -1.5569 -0.8552  1.6230  0.3552  0.4061
#> -2.8489  0.9953 -0.6664  0.8313  0.3048  0.0770 -0.6417  0.1305 -0.4681  0.4176
#>  2.3160 -0.0211  1.0990 -0.4006 -0.3004 -0.4107  1.1690  1.2653 -1.0676 -0.6067
#> -1.0153 -0.1934  0.0180  0.1404  1.2774  0.4408  0.9949  0.0367  0.7869  1.1244
#>  0.4930 -1.2405 -0.2605  1.0261  1.6136 -0.4481  1.1586 -0.5216 -1.2808  0.8724
#> -0.4476  0.3283  0.9831  0.4217 -0.0133  0.0536 -0.8177 -0.1369 -1.0696  1.7794
#>  1.3517 -1.2620 -0.2905  1.8818  2.1333 -0.2462 -0.8482 -0.0142 -0.4429  0.2953
#>  0.2782 -0.2847  0.3205 -2.8523 -0.5067 -0.0504  0.2453  1.0996  1.3279  0.0689
#>  1.3383 -0.2321 -0.5750  0.2385 -0.1691  1.5927  0.6094 -0.4697  1.3179 -1.0328
#>  0.1531 -0.8360  0.6447  0.0582 -0.0010  0.7195 -1.5267  1.2047  2.0948 -0.4929
#> -1.4958 -1.3048 -0.4513  0.8942  1.2925  0.2690 -0.5787 -1.1957 -0.3633  0.0886
#>  3.2058  0.1782 -3.3087 -0.8189  0.9399  0.5287 -1.7443  0.5805 -0.9216 -0.5463
#>  0.6942 -0.6033  0.8946  0.7329 -0.2486  0.2956 -0.1648  0.0659  0.5984  0.3244
#> -0.4700 -0.1261 -1.5958 -2.3085 -1.4482  1.3779  0.6793 -2.4290 -0.3148  0.7047
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]