Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.2809 -0.8839 -1.2899  0.7671 -0.3191 -0.1949 -1.0808 -0.4338  0.6623 -0.2422
#> -0.2880  0.6848  1.9671  0.2046  1.0775 -0.4769  0.5325  0.9351 -0.1692 -0.5280
#> -0.2146 -0.9669 -0.5702 -1.3853  0.7998 -1.9453 -1.6983 -0.4160 -0.6593  1.0168
#> -0.0741 -0.0929  0.2230 -0.3224  0.8509  0.5939  1.5146 -0.6222  1.0760 -0.5124
#>  0.2446 -0.5923  0.0002  0.8134  0.7300 -0.1926 -0.4565  0.1932 -1.0548  1.7544
#>  0.1218  0.2585 -1.0841  0.1050  1.4641 -1.3069 -0.5439 -0.6670  0.0187  0.8945
#>  0.3535  1.1685 -0.9206 -0.1349 -0.2008 -0.9996 -2.1572 -1.4682  0.8089  1.2208
#>  0.0839 -0.5747  0.7043  0.6244  0.0599  0.1498 -0.1841 -0.8273 -0.5536 -0.1297
#>  1.1800 -0.0788 -1.0721  0.1586 -2.3840  0.4904  0.0138  0.1186  0.0091  0.2919
#>  0.8046  0.4819  0.0581  0.6510 -0.2807 -1.0016  0.2303 -0.0412 -0.9346 -0.7479
#>  1.9459 -1.1744 -0.1102  1.6923 -0.7531  0.1101 -0.6141  1.0308  1.7831  2.8067
#> -0.9390 -0.7374  0.2751 -0.6211  0.3147  0.7911  0.9620  0.5097  0.7280  0.5068
#> -1.9143  1.8536  2.2452  0.7563 -0.4167 -0.5145  0.3068  0.0303 -0.7826 -1.1776
#>  0.0491  0.3538  0.1073  0.2839  1.0174 -2.3895 -0.2478  1.0429 -0.0921 -0.3816
#>  1.0287  1.9239  0.0886  0.9953  1.2610 -0.2053 -1.0669  0.6850 -0.8596 -0.3536
#> -1.3203  0.0634  1.4271 -0.1072 -0.6884 -0.3948 -0.2093  0.2305  0.3427 -0.8185
#> -0.4431 -0.6027  1.0502  0.0863  2.2892 -0.8141  1.4212  0.7119 -0.3107  0.4855
#> -0.8818  0.4751 -0.7387  0.1784 -0.4524 -0.3822 -1.0175 -1.4852  0.5363  0.7269
#>  1.8347 -0.9424  0.5198  0.1185  0.3143  1.4810  0.9427 -0.7941  0.0399 -1.8167
#> -0.5614 -3.1979 -1.0178  0.2333  0.0027 -0.4827 -0.8976  0.7324 -0.2398 -1.1315
#>  0.1948  0.6366 -0.2475 -0.1857  0.9558 -1.2671 -0.7953  0.4922 -0.5042  0.2979
#>  0.6658  0.1994  0.8023 -0.2585 -0.5416 -0.5376  0.6807 -1.3625 -1.4583 -0.0686
#>  0.5077  0.0496 -0.8320 -0.0326  0.0506  1.3549 -0.3265  2.0706 -0.6717  1.3788
#>  0.1889 -0.5787 -3.1272  1.1375 -0.7401  0.8134  0.2515 -2.2534 -0.1639 -0.3124
#>  0.5335 -1.2820  0.7847  1.1876 -0.6398  0.9496  0.1775  0.8504  0.4209  1.0933
#>  0.2829 -0.9481  0.1474  0.0567 -2.8301  2.0191  1.8249  0.3884  1.1874  0.3831
#>  0.4641  1.9186 -0.2715  0.0122 -0.6502  0.7066  1.0109 -0.7473 -0.5173 -0.2609
#>  0.4657 -0.7157 -0.4804 -0.5507 -0.2984 -2.1659 -0.4118  0.9477  1.0203  0.5247
#> -1.4412  0.1552 -1.3335 -0.7721 -0.2961 -1.2434  2.0353 -0.5334  0.3487  0.0725
#>  1.3517 -0.2062  0.7470 -1.6568 -1.8596 -0.4952  0.1938  1.1390 -0.5517 -0.4426
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]