Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6-4.8669e-01 -1.0313e+00 -1.6921e+00 -6.0923e-01 -1.1029e+00  4.9440e-01
#> -2.0005e-01  1.9558e+00 -7.2387e-02 -6.6265e-01 -1.7641e+00  1.1142e+00
#>  3.8965e-01  6.0405e-01 -5.7769e-02 -1.8965e-01 -1.2868e+00 -1.5311e+00
#> -9.8295e-01  6.9401e-01 -1.4230e+00  1.2107e+00 -1.1784e+00  6.8315e-01
#> -3.5896e-01 -4.5074e-02  1.0501e+00 -1.0398e+00 -2.2584e-01  6.4259e-01
#> -2.9442e-01 -2.6318e+00 -4.1298e-01  2.3629e-01 -4.3779e-01  3.7519e-01
#>  4.2713e-01 -1.4364e+00 -4.8560e-01  1.7163e-01  2.9826e-01 -1.6282e-01
#>  3.5575e-02  3.5833e-01 -8.4657e-01  4.9377e-01 -9.5743e-01  3.9026e-01
#> -1.8319e+00  3.9928e-01  2.1174e+00 -6.7749e-01 -7.7434e-01 -1.2788e+00
#>  8.2172e-01  1.7914e+00  8.0157e-01  1.2971e+00 -1.8370e+00 -1.2221e+00
#>  1.0883e+00  6.6973e-01 -1.2571e+00 -2.1612e-01 -7.2009e-02  1.3994e+00
#>  1.9527e-02 -1.0111e+00 -9.1069e-01 -2.7187e-01 -1.5396e-01  7.8578e-01
#>  1.6425e-01 -1.0643e+00  8.8541e-01 -8.7199e-01  1.7186e-01  5.9871e-01
#>  6.0464e-01  1.7708e+00  4.1100e-01  8.5683e-01  9.8047e-02  2.7468e-01
#> -8.3349e-01 -5.3545e-01  1.5653e+00 -1.0054e-01  1.7697e-01  1.9901e-01
#>  4.3529e-01 -9.3112e-01 -8.8418e-01  7.3627e-01 -5.7334e-01 -1.6972e-01
#> -1.9303e+00  2.1565e-01  4.6128e-01  3.4712e-02  1.6693e+00 -4.2259e-01
#> -4.4706e-02  5.4396e-01  6.8616e-01 -1.1633e+00  2.2274e+00 -1.1732e+00
#>  1.0094e+00 -6.6893e-01  2.6009e-01 -9.8919e-01  1.2911e+00 -5.3951e-01
#>  4.4259e-01 -1.5405e-01  1.1272e+00  1.5545e+00 -6.4745e-01 -1.6639e+00
#>  1.3109e-01 -1.4463e+00 -6.1024e-01 -1.9830e+00 -9.1405e-01  6.2813e-01
#> -4.3507e-02  7.0368e-01  8.4300e-01  2.0170e+00 -9.2451e-01 -9.2156e-01
#> -8.7410e-01 -8.0449e-01  2.6942e-01  6.5911e-01  6.5376e-01 -1.8261e-01
#>  1.1711e+00  3.8660e-02 -1.1037e+00 -6.0045e-02  3.6353e-02  1.6849e+00
#>  2.2699e+00  5.3420e-01  8.1145e-01  1.1240e+00  9.8660e-01  3.1471e-01
#> -8.4304e-01  3.7658e-01  1.7610e-01  1.7088e+00  1.1989e+00  4.4246e-02
#>  2.7924e-01  4.1556e-01 -1.4987e+00 -3.4288e-01  1.9957e+00  2.2087e-02
#>  7.4996e-01  2.5304e-01 -2.8801e-01  7.7528e-01  6.1747e-01  1.0063e-01
#>  3.4855e-01  5.0487e-01  5.2090e-01 -2.8070e-01  4.2255e-01 -1.7247e+00
#> -1.6595e+00  4.5941e-01 -5.8002e-01 -2.4018e-01 -4.3336e-02  2.6462e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]