Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6068  2.2169  2.0356  0.2073 -0.1179 -0.9961 -1.5948  1.6556 -1.5886 -0.2682
#>  2.1348  0.3722 -1.4376 -0.1709 -1.2353 -1.6964 -0.0143 -0.1933 -0.6071  1.3498
#> -0.4983 -0.9248 -0.0193  0.6412 -0.4295  1.0805 -1.7507 -0.0391 -0.2028  0.4263
#>  0.0190 -1.6126  0.1728  0.7894 -1.4318  1.1820  0.6886 -1.2871 -1.9379  1.1505
#>  0.9456  0.0589  0.9911 -0.1402  0.3574 -1.6873 -1.5321  0.1955 -0.6980 -1.9279
#>  0.7828  1.6697 -1.0900 -0.5466 -0.2866 -0.2883 -0.4359  1.1377 -0.4807  0.8135
#> -0.1056 -0.2883  0.7322 -0.1051 -1.3610 -1.1390  0.6236 -0.0196  0.4014  1.2795
#>  0.5178 -1.4094 -1.2360  0.7725 -0.8449  0.7964 -1.0598 -0.5549 -0.9364  1.3410
#>  2.0988  1.2413  1.5857 -0.0888 -1.1247  0.2240 -1.1045 -0.0586  0.3096  1.0379
#>  1.2052  0.2752  0.9085 -0.3923 -0.1633  1.1635 -0.5900  1.2582  1.2179 -0.1136
#> -0.2986 -0.9616  1.2513 -0.0231  1.1903 -1.0925  0.6048 -0.3661  0.0166 -1.7021
#> -0.1205 -3.0814 -0.9560  0.1130 -0.3120  1.3350 -0.3271 -0.2821 -1.3071 -1.0325
#> -0.3931  0.3101 -0.7770  0.1501  0.1006  0.1454  0.6981  1.1140 -0.4910  1.1154
#> -0.2454  0.2092 -1.4676 -0.4474  0.9995  0.4902  1.1277  0.6129 -1.3966  1.0479
#>  0.7923  0.9871  0.6798 -0.1939 -0.3282 -2.2789 -0.2128 -0.5906  1.5715 -0.6016
#> -0.5332  0.7169 -1.4976 -1.0382  0.8933  0.1604 -0.4226  0.3203 -1.4132 -0.8562
#>  0.4217  0.8437 -0.5338  0.8492 -3.1100  0.4911 -0.3107 -1.2684  1.1247 -0.1812
#>  1.6349  1.7400  0.4339  0.3692  0.3710 -0.1324  2.0193 -0.1017  2.1508  0.2731
#> -1.9563  0.7944 -1.1416  1.4923  0.8300  0.1589  0.1901  0.7217 -0.5081 -0.5457
#> -0.7011  0.1564 -0.2575 -2.7206  0.2946 -0.4146 -0.6186  0.8538 -1.0962  0.2390
#> -0.2056  0.4343  0.0707 -0.3329  0.6678  0.8256 -0.4640 -1.2587  0.5384 -0.8700
#>  0.2135  0.7388 -1.6619  0.8619  0.2187 -0.3636  0.0911 -1.9448  0.6191  1.3327
#> -0.2209  0.0805 -0.0289 -0.4488 -0.3872  1.4829 -1.7882 -0.5910 -0.5843 -0.9004
#>  1.0258  0.8354 -1.8986 -1.0668  1.9362  0.1685 -0.6365  0.1941  0.8419  0.5748
#> -2.1096  0.7844  0.3578  1.0971  1.0161 -0.0284  1.8880  0.9289 -1.5168 -0.0408
#>  0.2177  1.3080 -0.5119 -0.6204  0.1128 -0.7323 -0.3563  0.2055  1.3937 -0.2534
#> -0.0072  0.9894 -0.7738  2.0540  0.6175 -0.3930  1.0288 -1.0729  1.1660 -1.7851
#> -0.3117  0.1186  0.1210  1.9819 -0.8530 -0.5838 -0.7786  1.3053  0.1437 -0.4386
#>  1.0141  0.3430 -1.6034  0.9795 -1.5781  1.1483  0.8360  0.0043  1.1612  1.2191
#>  0.0325  2.1405 -1.0882 -0.2301  0.3315 -0.0124  0.6234  0.2117 -0.5589 -1.3365
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]