Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-0.0135 -0.6456  0.2980 -0.5025  0.3359 -0.3839 -0.6186  1.0244 -0.1139  1.0265
#>  0.4258 -0.1745  0.0412 -0.2195  0.1184  0.0080 -1.0523  0.2226  2.2632  1.3879
#>  0.1614  2.1297 -1.2634 -0.3028 -0.2645 -0.3977 -0.0130 -0.2778 -1.2822  0.6826
#>  0.0550 -0.0385  0.8192  1.1470 -2.8974  0.1523 -1.6372  0.1137 -0.1043 -0.0479
#> -0.3254 -0.6256  0.4492  0.0305 -0.8171  0.6394 -1.3101  0.3013  0.8057 -0.0360
#> -0.8186 -0.9912  0.3251  1.1267 -0.9402 -0.0260 -1.1381  0.3606  0.0547  0.1073
#> -0.7611  0.2747  1.1112 -1.3208 -0.2126 -0.0033 -1.0803  0.9487 -0.1961 -1.5335
#> -1.3873 -1.7944  1.1804  0.1594 -0.4811  0.0535 -0.7225  1.1947 -1.2983 -0.3196
#>  1.6170  0.4183 -0.8549  1.1304  0.4451  0.0576  1.5469 -1.7961 -0.2067 -0.1220
#>  0.1391 -0.2832  0.2494  0.0746  0.2171  1.9903 -0.5253  0.9934  0.7477  0.8957
#> -0.6241 -0.5793  0.8753  1.9769 -0.5298  0.0159 -1.7919 -2.0016  1.0265 -0.3980
#>  0.9618  0.1173  0.0764  0.3571  0.0488 -0.3646  0.2897  0.0205  1.1232 -1.6736
#>  1.2813  1.2121 -0.8537 -1.3289 -1.0825 -0.0899  0.2684 -0.4918 -0.4792 -0.5900
#>  0.4343  0.1836 -2.4439 -1.5936  0.1656  1.9124 -1.8265 -1.9078  0.4329 -1.4355
#> -1.3870  0.2061  0.7099  0.6682  0.4087  1.2559 -0.1627  0.3183 -0.3307  0.1639
#> -1.1811 -0.8921 -0.2947 -0.2142 -1.9364 -0.2772 -0.0847  0.5583  0.4293 -0.0640
#> -1.1841  0.2196  1.2516  0.5355  1.0087 -1.6900 -0.3463 -0.7495 -1.5414 -0.9345
#>  1.1693  1.7735 -0.0714  0.3750  0.5431 -0.8561  0.8602 -0.2008 -0.6514 -1.7397
#>  0.0893 -0.3765 -0.2009  0.4708  1.9997 -0.2826 -0.9673 -0.1753  0.6472  0.4222
#> -0.5392 -0.3062 -2.5164  0.8246 -0.0263  0.3038  2.0541 -0.0160 -0.2452 -1.2582
#> -0.4211 -0.5694  0.1105 -0.4092 -1.1820 -1.0045 -0.5641 -0.5540  0.0950  0.2957
#> -0.1775  0.8290  1.5337 -0.7190  0.9733  1.2042 -0.1469 -0.8251 -0.4203 -0.0382
#>  0.0928 -0.5572 -0.6893 -1.0653 -0.4861  0.6458 -1.7340  0.8166  1.9482  1.2721
#> -0.4484  0.7035 -1.4149 -1.4061 -0.1176 -0.4796  0.0676 -0.3374 -0.5726  0.6463
#> -1.1676 -0.1551 -0.1921  1.2438 -2.3246 -2.4011 -0.3899  0.0128 -2.2164  1.0378
#> -0.8951 -0.4247 -1.4395  0.2929 -0.9478 -1.8578  0.2107  0.4389 -1.0833  0.8403
#>  0.6367  1.2601  0.1723 -0.2150  0.0943 -0.8262 -0.0704  0.6836 -0.1140 -0.8729
#>  0.1448  0.6705 -1.8434  1.0510 -0.1805  0.7013 -0.6541 -0.0356 -0.5717 -0.4508
#>  0.9688 -1.0007 -0.1732  0.1229  0.4667  0.0149  0.0241 -1.0431  0.3532 -0.5222
#>  1.6684  0.9281  0.9900  0.0207 -0.0147 -0.7166  0.2704  2.2494  1.9533  1.3214
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]