Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0362 -0.1271  0.1169 -0.6555  0.1918 -0.0165 -0.7195  0.5581 -1.9694  0.4603
#> -1.6144  0.2103  0.1647  0.8954  0.5084  0.0314 -1.8724 -1.1815  2.0693 -0.3725
#> -1.6952  1.4542  0.9630  0.9469 -1.1524 -0.3607 -0.2415 -0.4400  1.0840  0.1850
#> -1.5025 -0.3816 -0.5591 -0.1653 -1.2491  1.4642 -0.9676 -0.8169 -0.9747 -0.3593
#>  1.5705 -0.6668  0.3672  0.2416  0.0916 -0.3097 -0.9204  1.0009 -0.6066  0.4150
#> -1.0009 -0.2702 -0.4110  0.6613  2.3681 -2.4741 -1.4893 -0.3810  0.6435  0.1380
#>  2.7656 -0.3209 -0.2733  0.1932 -0.0925 -1.1130  0.5469 -0.2290 -0.6095 -0.8985
#>  0.2372  0.1714  0.4931 -0.7708  2.5467  0.9240  0.4900 -0.7674  0.1213  1.7190
#> -1.4082 -0.7738  0.0280 -0.3897  0.7647 -1.1782 -0.8400 -1.7640 -0.4171 -1.1588
#>  1.1458  1.4372  2.7824 -1.4064 -1.1483 -1.1778 -1.0439  0.5833 -1.0851 -0.0364
#>  1.4108 -2.4292  0.9320 -0.5559  0.3278  0.3750 -0.3961 -1.0824  0.9733  0.7465
#> -1.0513  1.2044 -0.4031  0.8269 -1.7213 -0.7290  0.2054 -1.4852  2.2929  1.9643
#> -0.3646 -0.1272 -0.0061  0.7148  0.3442 -0.1360 -1.1411 -1.5909 -0.1714  0.7614
#>  0.0483 -0.2961  0.5984  0.0762  0.6393  0.1397 -0.5847 -0.1602  1.3225  0.4294
#> -1.4081  0.2571  0.5816 -0.6104  0.6716 -1.6951 -0.2235  1.2724 -1.2651  0.2189
#> -0.2638 -1.4274 -0.4920  0.1159  0.6415  0.3364  0.1235 -0.8740  0.5870  2.1814
#> -0.5488  1.0958  0.5175  0.3464  0.0372 -0.2431 -0.6416  0.9257 -1.4221 -0.4397
#> -1.5404  0.5892 -1.1218  0.6228 -0.6470 -0.5094  1.0072  0.3201 -0.6466  0.0054
#> -0.5706 -0.9915 -0.1782 -1.7894 -0.9590  0.2838  0.8692 -0.1565  0.0867 -0.7940
#> -1.0511 -1.5037  0.1466  0.8240  1.6638  0.6500  0.2041 -1.0963 -0.9314  1.6468
#> -0.8409 -0.4992 -1.7449  0.3113 -0.1556 -1.0911  1.5499 -0.0431 -1.2030 -0.0930
#>  0.7689 -0.2943 -0.4572  0.3189  0.5244  0.7629 -0.2978  1.0139  0.3123  0.8933
#>  0.4736  0.1452  0.4893 -2.0484  0.3888 -0.8786  1.8059 -0.7411 -0.7653 -0.4169
#> -0.1414  1.8851 -0.8199 -1.6145 -0.4102 -1.6336  0.6031 -1.2730  0.0010  0.0194
#> -1.1424 -0.6093  0.9731  0.6499 -1.5845 -0.9516  1.9086 -0.8627 -0.6209  0.3040
#>  0.8678  0.6583 -0.4730 -0.4467  0.7991 -2.1520 -0.0712  0.8212  0.6257 -0.0159
#>  0.1859 -0.7051  0.8273 -0.7554  0.5901  1.4330 -1.1410 -0.6648  0.6103  0.1705
#> -0.4783  0.8950  0.6755  1.2252  1.2885  1.1055 -1.1347 -0.7794 -0.7623 -0.4419
#> -0.1028  0.3002 -0.9729 -0.8056  1.4655  1.1233  0.6027  0.8435 -1.0458  0.6007
#> -0.0757 -0.7032 -0.3335  0.4828 -0.1563  1.2648  0.0640  0.0338 -0.9012  0.6998
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]