Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4173 -0.8157  0.4227  0.1640 -1.7694 -0.6228  0.8366 -0.8065 -0.1652  0.0529
#>  1.4742 -0.7539  0.3473  0.1098  0.1756  0.7441 -0.1326  1.2317  0.6986 -0.4898
#>  0.8735  0.0539  0.3745  0.1695  0.1141  0.1416 -0.6930 -0.9607  0.8946 -0.5362
#>  0.2539  0.1321 -0.1309  0.0746 -0.4084  0.9503 -0.2748  0.1385 -0.7035  1.9311
#> -0.6043 -2.4032  0.6434 -0.2432 -0.4447  0.3991 -1.0500  0.9270  1.0740 -0.8279
#> -1.8976  0.0392  0.7614 -0.4228 -0.6900 -0.1660  0.6582  0.2848  1.1486 -1.9997
#>  0.9851  1.3156  1.2271  0.8962  1.1881  0.8781 -0.8620 -1.2904  0.7674  1.2969
#>  0.8103 -1.3746  0.9198  1.2583  0.6921 -1.6024  0.9672 -0.0025  0.0977  0.2245
#>  0.9864 -1.6005 -0.5899 -0.1359  0.1683 -0.4284 -0.3122 -0.9225  0.0960  0.5197
#> -0.0570  1.2268  0.4590 -1.9954  1.5142 -1.1113  0.8281  0.5370 -2.5086  0.8403
#>  2.2607 -1.9818 -0.6481  1.3988  0.3349  0.8298  0.6072 -0.3700  0.9297  0.2648
#> -0.6263  0.7399  2.0397 -0.3902 -0.1442  0.0990  1.6686  2.5604 -1.4324 -0.0092
#>  0.9886  0.3665  0.0408  1.5332  1.5517  1.5212 -0.3411 -0.8004 -0.2509  0.8018
#>  2.1981  0.8112  0.8908  0.8981 -0.3128  0.4855  0.9578 -1.1225 -1.0331 -0.1295
#> -1.0607 -0.3939 -0.1220  1.3060  1.2429 -1.3224 -0.2514  0.0261 -2.0509 -0.7220
#> -1.0137  0.0634  1.2719 -0.0270  1.4088  0.4955 -0.5821 -1.1629  0.3200  0.7691
#> -2.3138  0.4864 -0.1292  0.6429  0.2105 -1.5113 -0.1326 -1.5788 -0.5042 -1.1572
#> -0.4334 -1.2640  0.9291 -1.1804  1.6542 -0.8423 -0.0248 -1.4433  1.7728 -0.2557
#> -0.8904 -0.4608 -0.4068 -0.5297 -0.3427  0.4906  0.6166  0.7734 -0.7933  0.7382
#> -0.9012  0.0029 -0.0602  1.1064  1.5604 -0.6793  2.0203 -0.0705 -0.5266  0.0279
#> -0.5418  1.8541 -1.1071 -1.2931  1.5817  0.0078  0.3139  1.4503  0.1306  0.1246
#>  0.4328  0.2971 -1.2586  0.0825 -1.0861  1.2272 -0.1453  0.4546  0.3600 -0.1731
#>  1.0352  0.4138  0.0282  0.6203 -0.0779  0.4016 -0.1097  0.3635 -1.0281 -0.1749
#> -1.1328 -1.1952  0.2915  1.3393  0.1503 -0.7019  0.6856 -0.3190 -0.9070 -1.0014
#>  0.1993 -0.5328 -0.3111 -0.7432  1.2279  0.4330 -0.8964  0.1383  0.0155  0.9616
#>  0.9268  0.3745 -0.3345 -0.2544  0.9219  1.1312 -0.2455  0.6619  2.0586  0.1119
#> -1.7704 -0.5155  1.0616  0.0314  0.6423  1.2293  0.6889 -0.7993  0.4062 -0.4337
#>  0.3356  1.5968 -0.1328 -0.1140  0.5165 -0.2877  1.7565 -0.8739  0.5922 -2.4568
#>  1.0224 -0.7529  1.3245  0.3302  0.5306  0.7513  1.3431  0.2959 -0.7232  0.5834
#>  0.7324 -2.1400  0.9714 -0.8279  0.3211 -0.9119  0.6866 -0.4176  0.0637 -0.3842
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]