Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.5347  0.1956 -2.5465  0.1013 -1.3940  1.5696 -1.0750 -1.4614 -1.3270 -1.3189
#>  0.4316 -1.9479 -1.9337  1.2045  0.2939  1.4214  2.2702  1.7718 -0.5717 -0.3408
#>  0.6427  0.5732 -0.6142  0.7928  2.1879  1.0898 -1.4601 -0.3523  0.0597  0.1651
#> -1.7579  0.4225  0.1469 -1.1460 -2.0280 -0.2060 -0.6711 -0.1248 -0.7896 -1.3774
#>  1.4072  0.1159 -0.2724 -0.9595 -0.5540 -0.0215 -0.0770 -1.4544  0.8717 -0.4250
#>  0.5125  0.2449 -0.1426 -0.7391  1.5180 -0.4642 -1.2863 -0.2789  0.9496  1.1203
#>  1.5716  0.1832  1.6636 -1.4895  0.0999  0.9556 -0.2693 -0.3229 -0.6519 -0.5477
#> -0.3577  0.5822  0.4246  0.2803 -0.3178 -0.7131  0.2496 -0.3450  0.7566 -0.8548
#>  0.3806  0.6961 -0.6412  0.3707 -0.1835 -0.9178 -0.0565 -0.2150 -0.7396  0.3625
#>  0.3618  1.3288 -0.1933  0.8012 -0.0017 -0.3964  0.9047  1.3940 -1.0191  0.3980
#>  0.5345  1.6990  1.3321 -0.1720 -0.3642 -0.9061 -0.7757  1.0104  0.7321 -1.3734
#> -0.4120 -1.7284 -0.2354 -0.5405 -0.0916  0.5361 -1.7037 -0.6771 -0.3799 -0.6303
#> -0.8577 -1.0743  0.6092  0.2983  1.2292 -0.0829 -1.0547  0.5700 -0.8451  0.5298
#> -0.0782  0.3046  1.0233  0.4305 -0.9177  1.5251 -1.2829  0.4717  0.3567 -0.1648
#>  0.3268 -2.2126  1.4967  0.9637 -0.2210  0.3207 -0.3939 -1.9126 -0.0022 -0.3449
#>  0.5311 -0.2701  0.3075  0.6986 -2.0070 -1.1447 -0.2662  2.2886  0.3417 -0.6278
#>  1.2020 -0.3901  0.3075  0.0777  1.1758  0.2826  1.2141 -2.5891 -1.3708  0.2548
#>  0.9241  1.3913  0.5763  1.2234 -0.2590  0.0831 -0.3822 -0.0747 -1.5845 -0.0846
#> -0.1914  1.4731 -0.0516 -1.4235  0.8299 -0.2107  0.0775 -0.7398  1.3724  0.2326
#>  1.0657 -1.2649 -0.9426  0.0008  0.5931 -0.1660  0.5045  0.9438 -0.0454 -0.4145
#>  1.4439 -0.6215  0.6481  0.4392  1.1609  0.2363  0.6120  1.4879  1.4316  0.4076
#> -1.0202 -0.6944 -2.5856 -0.3689 -1.0443  0.3746  1.7696  0.6925  0.2362  0.7084
#> -0.7451  1.3754 -0.2989 -0.1093 -0.2550  1.0341  0.2148 -0.4471 -0.6713  1.5640
#>  0.6714  0.0434 -0.4206  0.6699  1.0399  0.1782  0.7219 -0.6704 -0.6652  0.3585
#> -1.0839 -1.0959  2.1962 -2.3125 -0.6885 -1.9880 -0.2911 -0.1699  0.3785 -0.5316
#>  0.7326  1.2666 -0.6464 -1.5547  1.0496 -0.4053  0.4888 -1.7862 -0.1439 -0.5150
#> -0.2215  0.6165  0.6302  0.4368  0.6144  0.6822  1.6504  1.8533 -0.0656  1.1783
#> -0.1213 -0.3148  0.5200  0.3517 -1.4214  0.8418 -0.5182  0.8087  1.3589 -0.1360
#> -1.7804  1.7058 -1.2296 -2.3298 -1.6215  0.1872 -1.5485  1.1581  0.5742 -3.2475
#>  0.6325  0.7030  0.1144  1.0853 -1.1899 -0.4151  1.3576  1.9922 -2.0552 -0.7847
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]