Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10 0.1323  0.9655 -1.1567  0.2767 -1.2337 -0.1382  1.4983  0.2004  0.5580  1.0077
#> -1.6785 -0.3309  0.0169 -1.4604 -0.6256  0.2649  2.0280  2.1616  0.1516 -0.2972
#> -1.0273  0.7887  0.5381 -0.6912 -1.3051  1.8795 -0.6911 -0.8200 -1.0467 -1.4099
#> -0.5801 -0.0407 -0.6891  1.4955 -0.7915 -0.6214 -0.3463  2.0071  0.2161  0.9224
#> -0.6297  0.2719 -0.3722  0.5851 -0.3435  1.2930  1.0401  1.8909  0.9067 -1.4282
#> -0.7709 -0.3228 -0.6543 -0.2925  0.0233 -0.9065  0.3266  1.0734 -0.6532 -0.9539
#> -0.1751 -0.1042 -0.4438  0.1544  0.5352 -0.7519 -0.4648  0.1598  0.0431 -0.6416
#>  0.4941  1.7460 -1.3250 -0.3696  1.9139  1.2748  0.4600 -0.1742  0.1841 -0.1083
#>  1.6247  0.6070  1.1397 -0.1917 -0.3852 -0.3149 -0.9357  1.5197  0.3764  0.1118
#>  2.4771 -0.7551  0.4271  0.1014  0.0220 -0.1340 -0.1380 -0.1800 -0.3262 -0.6998
#> -0.4688  0.0265  1.1690  0.4226  0.1927  0.0805 -0.6377 -0.5090  0.3462 -1.4252
#> -2.2131 -1.1176 -1.7152 -0.8618  0.5676 -0.8527  1.8969  1.0142  1.8532 -0.8481
#> -0.4794 -2.2779  0.4543  0.0848 -0.0031 -0.1050  0.3313 -0.3344  2.2458 -1.0630
#>  0.6875  0.0097 -0.0093 -1.7537  1.6235  0.3015 -1.3155  0.0146 -0.6313  1.9201
#> -0.7627  0.4649  0.4539  0.0209  0.3115  0.7427 -0.4565  2.0636  0.7672 -1.5196
#>  2.7245 -0.3250  0.6227  0.5742 -0.5011  1.8292  0.4120  0.6482  0.5372  0.0453
#> -1.5484  1.0398 -0.6083 -2.0487 -1.1895  1.1797 -1.2270 -0.8640  1.4840  0.3999
#> -0.5433  1.8235  2.0946 -0.1350 -0.0068  0.5408 -1.3858  1.5743 -0.1980 -0.5365
#> -1.1787  0.2106  1.6021  1.3628  0.3190 -0.4914 -0.7145 -0.1334  0.7768  0.1875
#>  0.5807  0.5558  1.1804 -0.1441  2.0787  0.0524 -0.1795 -0.0567 -0.7504 -1.0149
#> -1.8976  0.2871  0.6760  0.1656 -0.0502 -0.1094 -0.5100 -1.2231  1.5308 -0.4650
#> -0.3575  0.8076 -0.0162  0.1186 -0.6337  0.8683 -0.0608  0.2052 -0.3247 -0.1230
#> -1.2356 -0.2259  0.0221  0.1184  1.2766  0.1480 -1.2141  1.0271 -0.1549 -0.1141
#> -0.3701  2.6986 -0.0190  1.0524  2.4300  0.7018  1.0545  0.9273 -0.1636 -0.6238
#> -1.2516  0.6138 -0.6095 -2.0100  1.7904  0.1656  0.0976 -0.3838  0.6712  0.4826
#> -0.7877 -0.7415  1.8105  0.0022  0.1690 -0.6857 -0.2158  1.7694  0.8376 -0.5690
#> -2.8065 -1.5994  0.5976 -1.3619 -0.3054 -0.5819 -1.1769  0.3051 -0.2526 -1.3600
#> -1.8066 -1.7480 -0.2322  0.3191  1.4772  0.6530  1.3858  0.8943 -0.4175  2.1620
#>  0.2931  0.6621 -0.1927 -0.3699 -0.4593  0.9256  0.6201  0.4735  0.7748  0.9310
#>  0.0716  0.1492  1.3559  0.3815  1.6655 -2.3116 -0.4121 -0.3827 -0.4261 -0.8637
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]