Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5395 -0.1846 -0.2396 -0.1418  0.4140  0.3650  0.4666 -0.8676  0.7286 -0.5370
#> -1.5839 -0.5275  0.0718  0.6934 -1.3111  1.3075  0.8161  1.4930 -0.9308  0.3707
#> -1.3592 -1.2234  0.6452 -0.5860  0.9907 -0.9154 -0.5920 -2.0847 -0.8985  0.3168
#>  0.4761  1.1604  0.3588  0.2223 -0.9756  0.2746 -0.1341  0.2601  0.3686  0.5781
#> -0.6327 -0.1131 -0.6157 -0.2474  0.1173  0.2088  1.6054 -1.9685 -0.3335  1.0204
#>  1.2101  0.3975  0.3174 -0.8104 -1.1927  1.3651 -1.2227  0.4893 -1.9532 -0.9648
#>  0.1430  1.3773 -0.0792  0.2117 -0.0134  0.8017  1.2967 -1.5883 -0.6156 -0.0988
#> -1.5940  0.5484 -0.1254 -0.7041 -1.2621 -1.1844 -0.1577  0.2918  0.0145  0.6017
#>  0.9717 -1.4894 -0.4362 -0.2573 -1.7698  1.1600  1.0854  0.5878  2.0451 -0.3367
#>  0.3989 -0.4614 -0.2841  0.7729  0.7489  0.5779 -1.1026 -0.3341  0.2212  0.6129
#>  1.6541 -1.3443 -0.1378 -1.1528  0.4202  0.8532  0.2957 -0.4970 -0.4128  0.0339
#>  0.8604 -1.9296 -0.8322  2.8304 -1.1556  0.1446  2.2376  1.2206 -0.3993  1.2818
#> -0.1171 -1.6623  1.1274  1.5650  0.5246 -0.9505  0.3454  2.0724  0.4663  1.2420
#>  0.9338 -2.0127 -0.5205 -0.3664 -0.2346 -3.0175  1.1018 -0.7066  0.4907  0.3038
#>  0.9841  0.8180 -0.3784 -0.4447  1.0956 -0.0811  0.4897 -1.1074  0.8954 -0.2141
#> -0.6732  1.4221  0.0123 -0.2318  0.0323 -2.3150  0.4492  0.0806  0.1593 -0.6380
#> -0.8182 -0.8928 -0.0508 -1.3445  0.6751  0.5787 -0.8900 -0.3088 -1.1328 -0.6259
#> -2.7111  1.4868 -1.2205  2.0224  0.4290 -1.0863  2.0800  0.3172  1.8512  0.4483
#> -0.2075  1.7368  1.2480  0.4525 -0.0482 -0.0731  1.1310  0.9973 -0.7677  0.6949
#>  1.5616  0.5223  0.9674 -0.1878 -0.1776 -0.5817  0.6591  0.1665  2.2841  1.1148
#> -0.0815 -1.0297 -0.6181  0.0416 -0.7323 -0.4651 -1.0027  1.3136  0.3192 -2.7110
#>  1.4177 -0.3710  0.9118 -1.3500  1.4950 -0.3638  0.2267 -0.6228 -0.1351 -0.2423
#>  0.2083  0.5342  0.0602 -0.3154  0.0451  1.3059 -0.7435 -0.2807 -2.1968  1.9614
#>  0.0164 -0.2499 -0.3142  0.1955 -0.3786 -0.0961 -0.7861 -0.2846 -1.2926  0.0791
#> -1.2958 -0.5184  0.6257 -0.8435 -0.5344 -1.1985 -0.1127  0.4494  0.0909 -0.6493
#> -0.2143  2.1502  1.0347  0.0610 -1.2367 -0.0377 -0.1302  0.6752  0.1052  0.8757
#> -0.2090  0.4947  1.5483  0.1040  0.5511 -1.0934  1.7788 -0.2726 -0.1067 -0.2164
#>  1.0102  1.0223 -0.8533  0.2090  1.3350 -1.7434  0.5267  1.2197  1.3522 -1.2925
#>  0.1260  0.9289  2.3078 -0.3369 -0.6849 -0.2691  0.2880 -0.8251 -0.4750  1.2344
#>  2.3346  0.2427 -0.2882  0.4160  1.8704  0.3224 -0.1505 -0.0892 -0.7794  0.9507
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]