Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4471  2.1732 -0.3345 -0.6218  1.8492 -0.5535 -0.6885 -0.0262  1.5260  0.0415
#>  0.0748 -1.1453  0.8604  0.6275 -0.4147 -0.3780 -1.6821 -0.9708  1.2143 -0.2868
#>  1.8639 -0.8220  0.9557  0.1033  0.6635 -0.6149  0.9882  1.0016 -1.2176  0.9762
#> -0.3881 -2.2619  0.2844  1.6284  0.9051  1.5465  0.1589 -0.6429 -0.9214 -1.1159
#>  1.6372  0.3734  0.2244 -0.2472  1.3703  0.9339  1.8115 -1.3885 -0.3538 -1.4165
#> -1.1955  0.2960 -0.0455 -0.9183 -0.7529 -1.1746  1.2738  0.3598  0.8601  0.8912
#> -0.6725 -0.8828  0.1472 -0.5120 -1.0999  0.2199 -0.0926  2.3870  1.2566  0.9073
#> -0.1743 -1.5548 -2.0562 -0.8354 -0.7976 -0.1625  0.7328 -0.4616  1.8245 -0.6305
#> -0.8481  0.6597 -0.6723  0.4929 -1.5505 -0.4173  1.1422 -0.7953  1.1960  0.3253
#>  1.0004 -0.8041  1.1989  2.4197  1.2505  0.0063  0.3950  1.3609 -0.5243  1.0762
#>  1.1416 -1.1425 -0.6345 -0.1404  0.3168 -1.1648 -0.3783  0.1562 -0.4507  0.1899
#>  2.1023  1.6405  1.1550 -0.1535 -0.3126 -0.1515 -0.7801  1.1960 -0.1843 -1.3739
#> -0.1611 -0.9289  0.4846 -0.2703  0.1194 -1.0674 -0.2302 -0.4600 -0.2891 -0.9870
#>  0.0713 -0.3803  0.1167  0.3159  1.2039  0.8680  0.2757 -1.1305  0.5709  0.4987
#>  1.4190  1.5103  1.1883 -0.7547  0.7615 -0.7806  1.1232  0.0811 -0.5795 -0.8166
#>  1.0991  0.0353 -1.3615 -1.2203 -1.7425  0.7811 -0.4590 -0.0338 -0.3112  2.8278
#> -1.0718  2.0266  2.0816  0.5022  0.3318  0.3208  0.9343  0.2369  1.9410  0.7089
#>  0.3534  1.3965 -0.1047  1.1991 -0.0250  2.0448 -0.6667  1.2821  1.0932  0.3427
#>  1.9793 -0.3575  1.2087  0.2734 -0.9380  1.5299  1.2490  0.0124 -0.9674  0.9579
#> -0.7407  0.2916  0.1530 -0.9313 -0.2577 -0.2236  0.9284 -0.4798 -0.2942  0.3088
#> -0.2895 -0.0084 -1.1002  0.2599 -1.1483  0.6435 -0.0376 -1.7506  2.4399 -0.6456
#> -1.1289 -1.8769 -0.1375 -0.3420  0.6116 -0.0134  0.7112 -1.1313  0.1160 -1.1306
#>  0.5778 -0.9574  0.0042 -1.4499  0.6215 -0.2892 -2.7177  1.0867  1.6681 -0.9407
#>  0.3954  1.9306  1.4267  0.0719 -0.0046  0.5606  0.3868 -0.7859  1.9672 -2.2752
#> -2.2936  1.7499  1.7254  0.9005 -2.2995  0.2343 -0.3754  1.1444 -0.8072 -1.1160
#>  0.4290 -2.1049 -0.1230 -1.0717 -1.3082  0.8979 -0.2491  0.3426  0.1568 -0.0182
#> -0.3779  1.1787 -0.3000  1.4625  1.0733  0.8590  1.3975 -0.0035 -1.1249  0.1534
#>  2.4765 -0.4234  0.5507  0.5268 -0.8131  1.4125 -0.0671  0.7299 -0.7723 -1.4145
#>  1.0072  1.1145  0.4275 -0.1955  1.4481 -1.4056 -0.0292  1.3429 -0.0572  0.9001
#> -0.7226 -0.7646 -0.3194 -0.2278  1.2527  0.7800 -1.1509  0.4953 -0.5067  1.5192
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]