Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.9191  0.1282 -1.4120 -1.0899 -2.6852 -0.4366 -0.2739 -0.1213 -1.3890 -0.3072
#> -0.1261  0.0377  0.4734  0.5345  0.5533 -1.1497 -0.5208  0.5968  0.8517  0.1426
#>  1.7265  0.2582 -0.4271 -1.0241  0.3667 -0.8708 -1.2469 -0.3245  0.4334  1.3856
#>  1.1194 -0.6064  1.6057  0.9023  2.0163  0.5610  0.2645  1.0052 -0.3997 -0.7123
#>  1.0386 -1.4362  1.5016 -0.7757  0.2585 -0.2259 -0.2109 -1.8723 -0.2460 -0.3442
#> -0.4995 -0.1131  0.1080 -0.0263 -0.2462  0.2845  0.3642 -2.4490  0.0870 -1.1947
#> -0.3174 -0.9507  0.0725 -2.2789 -1.2689 -1.4687 -1.5964 -0.1402  0.1272 -0.3244
#> -0.0133 -0.4744 -0.8755 -0.7028 -0.9093 -0.6857  0.0236  0.9670 -1.3662  0.2935
#>  1.1786  0.0179  0.6697 -0.2004  0.4886 -0.0616 -0.6260  0.3033  1.5331 -0.7191
#> -0.2426  1.4758 -0.5090  0.0782 -0.0241 -1.2426 -0.4885  0.8925 -1.8224  1.1165
#> -0.3527 -0.2505 -1.6700 -0.8620  0.1748 -0.0613 -1.3496  0.4052 -0.7261  0.2393
#>  0.9210 -0.6129  0.6297 -0.3188 -1.2652  1.2250  0.1664 -0.6689 -1.0528 -0.1970
#>  0.5496 -0.4912 -0.3663  0.2715 -0.7070  0.7734  0.5687 -1.3442  0.2560 -0.7666
#> -0.4383 -1.2328  0.3148 -0.2527 -0.2298  1.1335  0.5407 -0.8830  2.3470 -1.1259
#>  0.5413 -0.1060 -0.5909  0.4146  1.5560  1.0783 -1.5201  0.1818 -1.2213 -1.2179
#> -1.0659  0.6802 -0.2859  0.4708  1.5269  0.9747 -1.8241  0.5547 -0.1025  0.3592
#>  0.7422 -0.7796 -0.2100 -0.1791 -1.0725  0.2829 -0.5439 -0.5964  0.5762  0.1832
#> -0.4953 -1.4532  1.3833 -0.8207 -0.6092  1.2144  0.1890  0.4745 -0.1005 -0.1866
#> -0.6386  0.1503 -0.7246  1.3428 -3.6561 -1.0327  1.2200  0.4534 -1.4809 -0.8072
#>  0.7267 -1.0815 -0.3823  1.7382 -1.5443  0.6653 -0.3394 -0.9667 -0.0027  0.4699
#>  2.5829  2.7096 -0.2331  1.1853 -1.6245 -0.2492  0.7759 -0.0109  0.1092 -0.6531
#> -0.1830  0.5142  0.9068  0.2340  0.2873  1.0021 -1.5224 -1.0440 -0.5150  2.0029
#>  0.7424 -1.6576 -1.9433 -1.4044  1.4614 -0.2317  0.4588  1.7234  1.7773  0.3870
#>  0.4108  0.6492  0.2483 -1.4230  0.0124 -0.2168  0.3932 -1.2057 -0.9869 -0.2752
#>  1.7033 -0.2375  1.5901  0.5271 -0.6065  0.3131 -0.1113  1.7332 -0.2079  0.2848
#>  0.0611  0.2686  1.1196 -0.5255  0.8852  0.8975  0.5191 -0.1168  1.2503 -0.6982
#>  0.9026  0.1155 -0.7456 -0.8438  0.8244 -0.6005  1.4340  0.4642  1.2717 -1.3772
#>  0.6877 -0.1713 -1.8464  0.1737  0.2089  0.4780  0.1544 -1.3194  1.2427  1.2243
#> -2.0556 -0.1856 -0.3118  0.4391  1.6379  0.1013  0.4819 -0.8633 -1.3598 -2.5768
#>  1.2162 -0.2455  0.9245  0.5028  0.9607  0.0424  1.6334 -1.4148  0.1812 -0.0497
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]