Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 8.0431e-01 -1.3852e+00  6.2163e-01 -1.3664e+00 -1.0638e+00 -2.1628e+00
#> -5.6743e-01 -3.7122e-01  4.2551e-01 -1.7944e+00  1.0634e+00  4.2054e-01
#>  1.5327e+00 -3.2886e-01  2.3550e+00 -3.6949e-01  1.9814e+00 -1.4057e+00
#> -4.3454e-01 -6.7128e-01 -1.3657e+00  1.7694e+00 -2.0115e-01 -5.2512e-01
#>  7.0715e-01 -1.2126e-01 -2.0537e-01 -4.0288e-01 -1.2118e-02  1.2591e+00
#> -9.0308e-02  6.2442e-01 -6.1300e-01  2.0627e+00 -8.9656e-02  2.2743e+00
#>  1.0222e+00  4.8325e-02  1.7088e+00 -5.7959e-01  9.4967e-01  6.1640e-01
#>  1.7772e+00  7.0408e-01  1.0829e+00  1.6766e+00 -5.8412e-01 -2.4427e-01
#> -1.9940e-01  4.7683e-01  5.0649e-01 -2.2933e-01 -2.5563e-01  1.8147e+00
#>  3.8349e-01 -1.0125e+00 -9.0116e-01  1.9395e+00  1.0170e-01  1.9863e+00
#>  1.7169e+00 -6.8646e-01  1.0589e+00  2.8854e+00 -1.2471e+00  1.2263e+00
#>  1.5928e-01  1.4176e+00  1.1905e+00  2.3294e-01 -1.6642e+00 -7.6628e-02
#>  2.6597e-01  4.9430e-01  3.4181e-02 -1.7862e+00 -1.0134e+00 -3.8786e-01
#>  1.8828e+00 -4.5004e-01 -1.4704e+00 -9.0861e-01 -5.0422e-01 -5.3699e-01
#> -1.6451e+00 -1.7728e+00 -3.3238e+00 -1.0048e+00 -2.7338e-01 -1.9672e+00
#>  1.0473e+00 -7.8870e-01 -1.9097e+00  3.1665e-01  5.4195e-01 -1.1688e+00
#>  7.1156e-01 -8.8507e-01  6.4645e-02  1.6915e+00 -2.1258e-01  1.9626e-01
#> -4.8219e-01 -1.0888e+00 -3.6906e-01  1.3637e+00  6.0798e-01  2.7704e-01
#>  7.6698e-01 -6.4334e-01 -1.6646e+00 -1.6756e-01 -1.7151e+00  4.4601e-01
#> -6.7253e-01  1.0916e+00 -5.0057e-01  1.4500e+00 -5.3049e-01 -5.4436e-01
#>  9.3852e-01 -3.1850e-01 -7.6782e-01  7.3724e-01 -1.0746e+00 -4.6117e-01
#>  2.5321e+00  4.7986e-01  8.3828e-01 -7.0393e-01  5.7103e-01  1.8048e+00
#> -1.3532e+00 -1.4465e+00 -9.8540e-01 -7.7091e-01  2.9742e+00 -7.1016e-01
#> -3.6194e-01  1.1106e+00 -2.1757e+00  7.3062e-01  3.1622e-01  1.1395e+00
#>  1.1225e+00 -1.7010e+00  8.1592e-01  7.1320e-01  4.8043e-01  1.0300e+00
#>  1.1328e-01 -1.1503e-01  1.2424e-01  8.3825e-01 -5.8487e-01 -6.3726e-01
#> -1.5330e+00  6.5617e-01  6.8807e-01  3.1056e-01  2.1745e+00  3.6910e-01
#> -1.1226e+00  3.0576e-02  8.6200e-01 -5.2328e-01 -9.0217e-01 -6.5260e-01
#> -8.3593e-01 -1.3863e-01  2.7787e-01  3.7475e-01 -5.7970e-05 -9.7199e-01
#>  1.6086e+00  6.5107e-01  3.0276e+00 -1.2554e-01  2.6026e-01 -2.6919e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]