Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.9656 -1.0783  0.3473 -0.5895 -1.1293 -0.6472 -0.4122  1.3012  1.2722 -0.4249
#> -1.2898  0.5275 -0.4911 -0.3809  0.4712  0.0121 -0.7904 -0.4626  0.9801  0.9652
#> -0.4470 -2.1821  0.7965 -1.6987  0.3258  0.4952 -1.2214 -0.6903 -0.5714  0.9751
#> -1.4472  0.3973  0.5843  1.5846 -0.8736  1.0827  0.1487 -0.1235  0.5204  0.7735
#>  1.2413 -0.3263  0.4219  0.1714  0.3236  1.5874 -0.8957 -1.3174  0.5543  0.2473
#> -0.1801 -1.0739  2.3308 -1.9643  2.0121  1.6593 -2.6843 -1.1577  0.4055 -0.4078
#> -0.5382  0.6192 -1.6244  0.7324 -0.2785  1.0446 -0.1214 -1.0826 -1.2184 -1.5565
#>  0.9161 -1.0318 -0.9236  0.5707  1.7943  0.4979 -0.4879  0.5540  0.8840 -1.5953
#> -0.1500 -0.6115 -0.7503  0.6568  0.4829 -0.1393 -1.5207  1.5528  0.1464 -1.8458
#>  0.1845 -0.3715  0.2430  0.6581 -0.8743 -0.0802 -0.4964  0.4900  0.9009 -1.4114
#> -1.1856 -2.3930  1.5566  0.7131 -0.7860 -0.2717  0.9630 -0.7378 -1.0003 -3.0404
#>  1.1286 -0.0580  0.5909 -0.1860  0.1030  0.7361 -1.1387 -1.7932  1.2482  0.3529
#> -1.2746 -0.4772 -0.4052  0.7816  1.1917 -0.2380  0.7865  0.2379  1.5808  2.2446
#> -0.6782  0.4698 -0.0145 -0.5413 -0.1865  0.2844 -1.6487 -0.0478  1.7099 -0.6752
#> -0.8080  0.9544  0.3395 -0.5469 -0.0483 -1.0855  1.1382 -1.2765 -0.6070 -0.3425
#> -1.7800 -0.6130  0.8656  1.1942  0.4375 -0.4687 -0.5514 -0.7357 -0.0392 -2.1854
#>  0.4057 -1.2083 -1.1414  0.0816  0.4144  0.9153  1.1016  1.2724  0.4788 -0.7798
#> -1.3890 -0.8643 -1.5360 -0.3277 -0.4944 -0.4871  0.3105  0.3072 -0.7229 -0.2848
#> -0.1834 -0.4262 -1.7154 -0.6449  0.1698  1.5424 -0.6675  0.1978  1.0953 -0.8723
#> -1.0337 -0.9306  0.2505  0.1363 -0.9581 -2.5569 -0.5826  0.2033 -1.4295 -0.7416
#> -1.2308 -1.3408 -1.9036  0.6326 -1.2018  2.3986  1.3614  0.5203  0.3314 -1.0143
#> -0.0160  0.4406  0.6766  0.3190  0.5520 -0.3433 -1.6548  0.3366  0.4979 -0.8599
#>  0.4163  0.5334 -0.1559 -0.3566  0.3298 -2.6369 -1.2947  1.4520 -2.2500  0.6627
#> -0.1473 -1.4290 -0.2333 -0.4481  0.7463 -0.9551 -0.2208  0.2747 -0.8052  0.6108
#>  0.7228  0.0875  1.2261 -0.9632 -0.5138 -0.8490  0.9321  0.2878  1.3619  0.4782
#>  0.4598  1.7509  0.9956 -0.6734 -2.4352  1.5206  2.1972  2.0548  1.8318  1.4922
#>  0.3245 -1.1232  0.1889 -0.1826  0.3227 -0.4153 -1.0304 -1.4376  0.7669 -0.8444
#>  0.2252 -0.4503 -1.2001  0.8216  1.6748 -1.0004 -0.9189 -1.8897 -0.4994  0.2759
#>  0.2702 -1.2326 -0.4206 -0.0575 -0.0266  0.1758 -1.2177  0.3875 -1.5548  0.5998
#>  0.7384  0.8252  1.1945 -0.4382  0.2820  1.4972  0.3263  1.1735  1.6688 -0.0401
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]