Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.8717  1.0289  1.1834  0.1547 -0.3018  0.3352 -1.1554  0.8725 -0.7247  0.3808
#>  0.4234  1.9901  1.4683 -0.3129  0.9544  0.2209  0.8483  1.0349  0.4150 -0.0345
#> -0.8397  0.2456 -0.0347  0.3837  0.6235 -0.5658  1.9036 -1.0080 -0.6184  0.4417
#>  0.2661  0.9587 -1.7488  0.2091 -0.1748 -1.7208 -0.5637  0.4541  1.8264 -0.1521
#> -0.3870 -0.7296  0.0009  0.9893  1.2611 -0.8603  1.7430 -1.0065  0.4464  0.5707
#> -0.0315 -1.6011  1.0633 -0.0862 -0.6491  0.1994 -0.5950 -0.2126 -0.0380 -1.9621
#>  0.3939 -0.6954  0.9366  0.2219  0.6544 -1.2552  0.1664  0.5110  0.6265  0.0168
#>  0.2032  0.3681 -0.4505  0.4067  1.3916 -1.1227  1.1372 -1.0186  0.5168  0.8531
#> -0.2614 -0.5760 -0.7916  0.5491  0.6956  0.4052  1.5990  0.0921 -1.4484  1.3554
#>  0.7654  0.2501  0.3653  0.8293  1.2626 -1.2219  1.7562  0.3454  0.9539 -2.6453
#>  1.1736  1.1305  1.1467 -0.5604 -1.2339  0.0929  1.7658  0.0373  0.6055  1.1665
#>  1.2756  2.2467  0.4851  0.9424 -2.1521  0.0532 -1.0209 -1.6482  1.0493 -0.1844
#> -0.5572 -0.1248  1.0622  0.3462 -1.5600  0.1603 -2.5411 -0.0008 -0.6995 -0.4621
#> -1.3657  0.0662  0.4694 -0.0639  0.0702 -0.1251  0.7103 -0.2669 -0.3491  0.4001
#>  0.5920 -1.1032 -0.7566 -0.3113  0.1253  0.1641 -0.6580  0.6411  1.0121 -0.0156
#>  0.1922 -0.6293  0.1587  1.2412 -0.7813 -1.0304  1.4653 -2.3127  0.2216 -0.4720
#>  0.4232  0.1513 -0.3092 -1.0241  0.2411 -0.3739 -0.5561 -0.0219 -0.2017 -0.3510
#> -2.4407 -0.2811 -1.0030  1.0775 -0.3276  0.9609 -0.6522 -0.1303 -0.9576 -1.1969
#>  0.4165 -0.9687 -0.1385 -1.1802 -0.2277 -0.3309  0.6301  1.0339 -0.6107  0.1039
#> -1.5506  0.8393 -0.4768 -0.2380 -0.9690 -0.6429  3.1029  0.0825 -0.0104  1.3443
#> -0.7619  0.0370 -0.4045  1.7704  0.0637  0.7759 -1.0441  0.7501  0.3659  1.3360
#>  1.2068 -1.3306 -0.1005 -0.2932 -1.4763  0.1667 -0.2668 -0.8824  0.0554 -0.8774
#> -0.3720 -0.0334  1.0047 -0.6774 -0.7082 -0.3494 -0.0813 -1.1575  1.2740  0.4539
#> -1.5236 -0.4383 -2.2211 -1.7504 -2.1944  0.0975 -0.0055 -0.4927 -1.1564  0.2532
#> -1.1337  1.5069  0.0245  1.4245  0.8960  0.4195  1.3202 -1.1416  0.2532  1.1979
#> -0.2703 -1.3407  0.7363  0.1130  0.9227  0.7032 -1.6042  0.0703 -0.0021 -1.0192
#> -0.1572  0.2306 -0.3864 -0.0126  0.4730 -0.6317 -0.1141  1.3338 -0.3436 -0.2738
#>  0.1233  0.3741 -0.1974  0.2520 -0.9726  0.5015  0.6040 -0.0534  1.4666  1.0980
#> -2.7109  0.4560  0.2687  1.5025 -0.6238 -0.5148  1.4966 -0.2295  0.3142 -0.4709
#> -0.1428  0.1170 -0.2554  0.4069 -1.5532  1.2047 -0.0290 -0.5305 -1.3349 -0.0924
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]