Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.4744 -0.9208  0.4137 -0.1101 -0.8231 -0.7543  0.3068 -1.0034 -0.7703  0.6097
#> -0.1204  1.8104 -0.3345 -0.5698  0.9274 -0.9546 -0.0643 -0.6277 -0.7205 -0.7564
#>  0.9928 -0.8033 -0.4322 -1.1424 -1.0599  0.1238  1.0976  2.1323 -1.2849 -0.2620
#> -1.5810 -0.2564  0.6556 -0.2092 -1.3771  0.4264 -1.1872 -0.2735 -1.3716  0.8865
#> -0.1868 -0.4482  0.8333  0.3536 -0.6170  2.7904  0.2275 -0.5396  1.4856  0.5669
#>  1.9954 -0.3186 -0.4479  0.5271  1.9622  0.9910 -1.2872  2.2469  0.2217 -0.5919
#>  1.2867 -0.5684  0.9862 -0.1493  1.5394 -0.1800 -0.2475  1.3445  0.1801  0.0737
#>  0.7547 -0.2722  0.8522 -0.1577  0.5792  1.5264 -0.4325 -0.1650  0.2557  0.2875
#> -0.3311 -0.5296  1.0983 -1.1469 -0.1777  1.4426 -0.4256 -1.6395  0.4281 -0.5704
#>  1.1246 -1.6408  0.8948  0.3318 -0.2809 -1.6613 -0.0946 -0.6884 -1.4014  0.2135
#>  0.0061 -0.8849  1.8088  0.7647 -0.5841 -0.8501 -1.4048  0.1764  1.4011  0.4059
#>  0.3903 -1.4977  0.1825  0.2285 -0.2524 -0.2690  1.3182 -0.8418 -0.0918 -0.0062
#> -0.4621  0.4699  0.8719 -0.4806 -0.7267  0.5164 -1.2278 -0.4834 -1.0275  2.2140
#> -0.8583 -1.8990 -0.1119 -1.2058  0.2737  0.2113 -0.7411  1.0658  1.1888 -0.3202
#>  0.3018 -1.6062 -0.6423  1.8634 -0.7817  0.9374  1.0672 -1.5214  0.7210 -0.2790
#>  0.1442  1.4790 -0.0078 -1.0557 -0.2467  0.0152 -0.1642  0.8295  0.8842 -1.0839
#> -0.7818  0.4260 -0.9435  0.4112 -0.6429 -1.3265  1.0014 -1.2689  1.1865 -0.5971
#>  0.0786  0.3076  1.0508  1.1261 -0.8982  0.1408 -0.8061 -1.2513  0.2763  1.1753
#>  0.8620 -0.8168 -0.4604 -1.5208  0.1771  0.6520 -0.3660  1.9060  0.4991  0.5871
#> -1.6022  1.9894 -0.6274 -1.1294  0.8899  0.5970 -0.2621  0.2757  1.6294  1.1479
#> -1.1160 -0.1857 -1.0262  0.5905 -1.7777 -1.1524 -1.1280  0.0889 -0.2019 -0.8495
#>  2.5203 -0.0446  0.2722  1.0898  1.5420 -0.0183  0.8375  1.2465  0.2103  0.1808
#>  0.1055 -0.0745  0.6329  2.6562  1.0913  0.3821 -0.3353 -0.0222  0.5714  0.7042
#> -0.1554 -0.6570 -1.2227  0.9164 -0.0627 -0.8692  0.3754  0.2562 -0.2229 -0.1816
#>  0.0388 -0.8669 -0.6098 -0.1415 -1.9032 -1.6623 -0.1061 -0.2183 -0.5768  0.9102
#>  0.1295  1.4244 -0.6640  0.4490 -0.1366 -0.9082  1.1706 -0.7431  0.1915 -0.7598
#>  0.8114 -0.8592  0.2779 -0.9734 -0.3639 -0.5805  1.8771 -1.3750 -0.3846 -0.2295
#> -0.9189  0.2841  0.1428  0.3676 -0.8461 -1.0324  0.4354  0.5519  0.4894 -0.3743
#> -0.7231  2.1487 -0.5375 -0.2920 -0.1596 -1.4924 -0.0544 -0.2344 -0.7841  0.2229
#> -0.7870 -2.5938 -0.8726 -0.3871 -1.0236  1.6191 -1.4835  0.7441  0.6339 -0.3841
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]