Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-2.3181 -0.3238 -0.8816  0.9979  0.3548 -0.3206 -0.6195 -0.0114  1.1444 -0.6798
#>  0.5326  1.6191  0.9752 -0.2656 -1.2957  0.5977 -1.4624  0.3170  0.1500  0.4375
#> -1.4170 -0.1546  0.4681  1.5375 -1.8075 -0.8442 -2.3240  1.5354  0.9853  0.9010
#> -1.6611 -0.2958 -0.4351  0.9011  0.4130 -0.9415  3.0499 -0.7205  0.0612 -0.9126
#> -0.5159 -0.2464  0.7234  2.1194 -0.5396 -0.4772 -0.1686 -2.6211 -1.6202  0.7035
#> -0.4838  0.6114  0.8817 -0.7812  0.2904  1.8659  1.9670 -0.5036  0.0163 -0.6785
#> -1.5007 -1.6673  2.2189 -1.0570 -2.7485 -2.5741  0.5752  1.0122  0.4204  1.1490
#> -0.1858  0.2747 -0.9037  0.9871  1.4044 -0.0514  1.3980 -1.2508 -0.4062 -0.5319
#>  1.1221 -0.8697  1.5050 -0.1039 -1.2657  0.3079  0.0217 -0.4090  0.7201 -0.8055
#>  0.7901 -0.3847 -1.8103 -2.4408 -0.1875  0.5156 -0.1477  0.5914  1.5465 -0.7032
#>  0.4496 -0.1743  0.5156  0.5052 -0.1496 -1.0067  1.2952  0.2637 -0.3013  0.1764
#>  1.1875  1.5686  0.4018 -0.1923 -0.8651 -0.4516  0.8478  0.4517  0.2714 -0.3814
#>  1.5943  0.0837 -1.4376 -0.6291  0.1612 -0.5774  0.7738  0.2478 -0.3180 -1.4277
#>  0.8010 -0.0396 -0.9157 -0.6960  0.0184 -1.2243 -2.1673  1.4670  0.7398  1.2712
#> -0.7743 -0.9850  0.6590 -2.3187 -0.8887 -0.0468  1.0493  0.0095 -0.6249 -1.4113
#>  0.2470  0.9618 -1.3423 -0.2424  0.8774  0.0484  0.4873  1.3098  0.3021 -1.2101
#>  0.3458  1.2415 -0.9154  0.9272 -1.9100 -0.0135 -0.8191 -0.5783 -1.3668  1.0045
#>  1.2148 -0.0379 -1.7454 -0.4088  0.3777  0.8241 -0.6367  0.3618 -0.9945  3.2163
#> -0.2454  0.2642  0.7436  0.6437 -0.7173 -0.7421  0.3473  0.2618 -1.1158  1.8374
#> -0.2424  2.8427 -0.1389  0.5637  0.4220  0.4253  0.2293  0.8043 -0.4269  0.3268
#>  1.3180  1.5012 -1.0272 -2.3637 -0.6154 -0.3951  0.5541  0.5172 -0.4275  0.5795
#> -0.0894  0.0821 -0.0988  1.2718 -0.2143 -2.0311 -1.1132 -0.2730 -0.7931  1.3047
#>  1.3625 -0.4040 -0.1457 -0.0536  0.6268 -1.1333 -0.5783 -1.0153  0.0164  0.3830
#>  0.3408 -0.8353 -1.8494 -0.5495 -0.2634 -1.1977 -0.6112  0.1952  1.0055  0.4236
#>  0.7045  0.9697  1.1948 -0.6772  2.4672  0.3600  0.0891 -0.0763 -0.1429  0.4952
#> -1.5461 -0.1179 -0.2070 -0.4615  0.9303  0.8091  0.1914  1.8029 -1.5036 -0.7354
#>  0.0351  0.2609  0.2181 -0.6383  0.5091 -0.6088  0.6600  0.8653  1.1592 -0.9627
#> -0.2466  2.0310  0.2162 -1.3799  0.9792 -2.2442 -1.0504  0.7151 -1.2102  0.0302
#>  1.1329 -0.5206  1.9280 -1.4350  0.9560  0.9049  0.2271  1.1068 -0.5532  2.0280
#>  0.3430 -0.4787 -0.9478 -0.8854 -0.8183  0.9925  0.7761 -0.5002  0.1930  0.9319
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]