Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.8699 -0.6934 -0.1790 -0.9365  0.6798  1.1427  0.2314 -0.0300 -0.2929  1.1385
#> -0.2680  0.2958 -0.8782  0.4526 -0.4365  0.7540  0.7674  0.4947  0.6752 -1.4650
#> -0.4875  2.3213  1.0740 -0.5653  0.4138  0.1992 -1.4456  1.9118 -0.4550 -2.1345
#> -0.9932 -0.5524  1.0503 -1.8207 -0.0469  1.0672  0.3512  0.4956 -0.9828 -0.1337
#> -1.0678 -0.4187 -0.0403 -1.4170 -0.3248 -0.5642  0.4161 -1.4810  0.2670 -1.8867
#>  0.9549 -0.6885 -0.9934  1.4249 -1.5236 -0.6908 -1.7886 -1.3250  0.0932 -0.2332
#> -1.6351 -1.3362 -0.1545  1.7385  0.7691 -0.5669  2.9211 -0.5454 -0.2910  0.4105
#>  1.0161  1.4975  0.4582  0.4620 -0.7240 -0.2324 -0.6721 -0.5026 -2.2691  0.4580
#>  1.8834  2.2955  0.5766 -0.5640 -0.7111  2.2048  0.7639 -0.9763 -0.0252 -0.0953
#> -1.8545  0.5228 -0.7018  0.9208  0.5856  0.5645 -1.0610 -0.5261 -0.1554  0.0899
#>  1.3516  0.6221  0.2055  0.0346  1.4698 -0.8407  1.1188 -0.9670 -0.4686 -0.0246
#>  0.1616 -1.6302  1.8119  0.0348  1.3869 -0.8806 -0.7825  1.2956  0.4953 -1.0143
#>  0.0179 -1.0985 -1.0161  0.6222 -2.0932 -1.7249 -0.8489 -1.1133  0.0801  0.2354
#>  0.1019  1.2112  0.1609  2.9105  0.9785 -0.5882 -1.1349  0.6401 -1.4276 -1.3013
#>  0.0042 -0.3827 -2.2442 -2.5068  0.2986  1.2538  0.1458 -0.7689 -0.1002  0.9622
#>  0.1849  0.7482  0.3834  0.0099  0.6708 -0.5194 -0.1513  1.2790  0.3551 -1.0272
#> -0.4584 -1.6575 -1.2935 -0.2782 -1.5980 -0.1185  0.2845  0.5341  2.0153 -0.9432
#>  0.1198  0.9337 -0.2962 -0.3159  1.0269 -0.9695 -0.9521 -0.1628 -0.1240  0.5593
#>  0.5234  0.5391  0.8103 -0.0035  0.6084 -0.3068 -1.1910  1.3698  0.7787  1.5154
#> -0.3968  3.1737  0.8277  0.8341 -2.3608  0.5817 -0.0785 -0.6728 -2.2394  0.8015
#>  1.8746 -0.1270 -1.3366 -0.4154 -0.4473 -0.4556 -0.8265  0.9956  0.0776 -0.0471
#> -1.4262 -0.3051  0.1167 -0.2430  1.0259 -0.1890 -1.8782 -0.0949 -0.4161 -0.0825
#> -1.9338  0.4819 -0.3093 -0.5608 -0.2138  0.4309 -0.1890  0.6190  0.4813 -0.6094
#>  0.2337 -0.6936 -0.2094  0.2509  1.3128 -1.8089 -0.8285  0.9709  0.1716 -1.1369
#>  0.9839  0.9937  1.1777  1.9217 -0.6987  1.4469 -0.6324  1.3655  0.7722 -0.5901
#> -0.5289 -1.1958  0.4058  0.9138  0.4007 -1.2436 -0.1045  0.0970  0.6519 -0.5110
#>  0.9465 -0.3060  0.3772 -0.1307 -0.8313  0.6524  0.2532 -1.0771 -0.3996  0.5440
#>  0.5931  1.2807 -1.6589 -1.6942  0.5731 -0.2578 -0.0833 -2.7083  1.5006 -0.4574
#>  0.7344  0.3165  1.1128 -0.0991  0.4355  0.7731 -2.3452  0.3354 -1.0691  0.1425
#>  1.2501 -0.7968 -0.4661 -0.0209 -0.6414 -0.3332 -0.9217  1.6807 -0.2897 -0.2616
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]