Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.9139  1.0718  0.4414  0.9923 -1.0196  0.6066  1.2035  1.0033 -0.6105 -1.1605
#>  0.8449  0.2759 -0.7707 -0.7802 -0.0480 -0.7212 -1.3448  2.0281  0.8987  0.5351
#> -0.4086 -0.7959  0.7173 -0.7799  1.1285  0.5255 -0.0172 -0.5859  0.5588 -1.3237
#> -1.6618  2.4998 -0.1981  2.0528  0.3108 -0.4430  1.4165 -1.7660 -0.5180  1.9069
#> -0.1352 -2.3951  0.3826  0.0486 -1.1308  0.4904  0.4718  1.7753 -0.2080 -0.6662
#>  1.4173  1.2850 -0.4514  1.4550  0.1570  1.1976  1.5388 -0.8923 -0.7712 -0.9627
#>  0.7378 -0.5670 -1.1940 -0.6138  1.4513 -0.3511  1.5764  3.6870  0.5220  1.8936
#> -1.0708 -0.2402 -1.1515  0.1970  0.3403 -1.1374 -2.0828 -0.5882  1.2133  0.8128
#>  0.0270  0.0065  0.5639  0.2631 -0.0958  0.3883 -0.2071 -0.1769 -0.7928  0.6125
#> -0.7491  1.0316  0.8291 -0.1067 -0.0452  0.2674  1.3601  0.8818 -0.1997  0.7408
#>  0.5253 -1.5614 -0.8577 -0.2747  0.4652  0.3928 -1.6570 -0.9412 -2.0594  0.4169
#>  0.4708 -0.2655  0.3310  0.1460  0.2730 -0.9798 -1.2209 -1.3341  0.0375 -1.0007
#> -0.5388 -0.4285 -0.1767 -2.0996 -1.2416 -1.1456 -2.4647  0.5783  0.2063  1.3800
#> -0.6102 -0.7088 -0.1261  1.7055  0.6867  0.1551  0.5723  0.6868 -0.3588  0.1305
#> -0.4885 -0.7436  0.2513  0.5177 -0.7122  1.5885 -0.1982 -0.6260 -0.4893  0.4784
#>  1.1477 -0.3992 -0.4125  0.8535 -0.6777 -0.5745  0.0025  0.3505  1.2444 -1.0860
#> -0.0136 -1.3582  0.0416  0.0370 -0.4706 -1.0065 -0.9171  1.1222  0.2334  0.2724
#>  0.4662  0.3166 -0.5088 -0.4995  2.1474  0.1514 -0.4472 -0.4853  0.0885  0.9872
#> -0.7404  0.6538  0.6105 -0.0346  0.8545 -0.5076  1.7241 -1.5517  1.3568 -1.4730
#> -0.7812 -1.1815 -0.7080  0.6824 -0.3664  0.2851  1.5978  1.0626  2.2956 -1.5784
#> -1.9630  0.8067 -0.1555  0.1672 -0.5231  1.0336  0.7383  0.0759  1.0192 -1.4737
#>  0.3999 -1.2857  0.1102  0.1042  1.6421 -0.0965 -0.3343 -1.3636 -0.9650  0.1102
#>  0.5781 -1.6518  1.0981 -1.4842  0.1362 -0.1969  0.4939  0.6326 -0.3873 -1.7481
#>  0.3566 -0.4550 -0.6825 -1.6445  1.8550  0.4180  0.5762 -0.7415 -0.4539 -1.1428
#>  1.9903  0.7898  1.4495 -0.0190 -1.5478 -0.7872 -0.9535 -0.9179  0.4529 -0.1659
#> -0.5616  0.5380 -0.1726  0.1090 -1.4073 -0.5589  1.4753 -2.2400  0.1091  0.3909
#>  0.2024  1.8022  0.6478 -1.3087  0.2632 -0.0295 -0.1402  0.6157  2.0225 -0.4636
#>  0.5570 -2.4618 -1.3282 -0.4276  0.6909 -1.2604 -0.4545  2.3118  0.4261  0.1887
#>  2.0588 -1.2451  0.0235  0.5949  0.5492  0.2943 -1.1189  0.9073  0.0701 -0.2477
#>  0.4067 -0.8828  2.2482  0.3052 -0.6954 -0.4609 -1.5509  1.6372 -0.0510 -0.7955
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]