Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.4437 -0.0442 -0.8572  0.9373 -1.0541  0.5906  0.9391  0.0405  0.1612  1.2406
#>  0.6355  0.1962 -0.7818  0.0799 -1.8553  2.1147 -0.5657  0.0477 -0.5796  0.5206
#>  0.3933  1.3823  0.3549 -0.3515 -0.1511 -0.2398 -0.7025 -1.1904  1.3114  1.7858
#>  0.8476  0.4202 -0.5945  0.0997  0.9173 -0.2589 -0.8149 -0.7588  1.2949 -0.1548
#> -0.2325  0.6059  0.2707  0.0154  1.2262 -0.1190  0.6228 -0.7341 -0.0889 -0.0983
#> -1.2187 -0.6170  1.1384 -1.4819  0.3164  0.1441  0.8923 -0.8197  0.3630  1.0985
#> -1.2032 -2.0307 -1.4342 -0.8502 -1.1923 -0.4553  0.8604 -0.1147  0.1114  0.3759
#> -0.4231  0.4994 -1.0290  0.1498  1.6812 -1.0265  0.0511  0.6735 -1.0688 -1.6610
#>  0.5422  0.0407 -1.3922 -0.0134 -0.6181  0.1810  0.4573  0.3299  0.9429  0.6618
#> -0.0416  0.6621  1.5080 -0.8741  0.5352 -1.1196 -0.6954 -1.2401  1.8205 -1.2437
#>  0.0524  1.4194 -1.2242 -1.0836  0.4951  1.0297  0.0045  0.8744  0.8637 -2.0956
#> -1.0666  1.8325 -1.0916 -1.4208 -0.4148 -0.3796  0.3463  1.4306 -1.3883  0.2992
#> -0.2298  0.1388 -0.6313 -0.0321  0.4356  1.6398 -1.4189  1.2680  0.5949  1.2929
#> -1.9693  1.5765 -0.0092 -0.6664 -0.6164  0.5520 -0.2481  0.6960  0.2722 -1.3609
#>  0.2260 -0.0114  0.8986  0.2443 -0.9552 -0.1843  0.3440 -0.6180 -0.8167 -0.0646
#>  0.0867  1.0460  0.1216  0.4547  0.2867 -0.2768 -0.3209 -0.3560  0.3105 -0.7683
#> -0.0069  1.2023  1.0308 -0.2745 -0.1132 -1.6688  1.1944  1.5411  0.1934 -0.1677
#>  0.7653 -1.2214 -0.9727 -0.4194 -0.3113 -0.9391  0.6795  0.4332  1.2714 -1.0943
#>  1.9340  0.6939  0.3177  0.9389 -0.7183  1.3748  1.1021 -0.8669  0.6517 -0.0920
#> -0.7332  1.0770  0.1504 -1.0368 -0.9811 -1.0705  0.8799 -1.0599 -0.7733  0.9549
#> -1.8784  1.9538 -0.1497 -0.3085  0.9804  1.1261  1.1606 -1.1990 -0.4578  0.5612
#> -1.5598  1.4144 -0.6979  2.9412  0.6148  0.4548 -0.6928  0.6122 -0.4392 -0.5300
#>  0.1135 -0.5959 -0.0556  0.9492 -0.1069 -0.4512  0.3136 -1.6037 -1.6186  0.0078
#> -1.6552  0.1804 -0.0267 -1.9410 -0.2362  1.2619  0.3533 -0.0293 -2.4792 -1.3328
#>  0.7553  0.8828 -0.5251 -0.6994 -1.6350 -0.6257 -0.7306  1.3671  0.7845  0.8433
#>  1.0877  0.3452 -0.2141  0.1881 -1.7660  0.1271  1.4340  1.0070 -2.0119  0.3194
#>  1.0465 -1.2301  1.6094 -0.1311 -2.3362 -0.3506 -0.2189 -1.0339  2.2760 -0.8267
#>  0.2334  2.3236  0.3736 -0.4869 -0.6628 -0.0413 -0.9509 -0.2132  2.6105 -0.7124
#>  0.8570  0.0677  2.0460 -0.6393  0.7207  0.5835  1.0200 -0.8879 -1.3536  1.3212
#>  0.1967 -0.6956  0.5416 -1.1702 -0.3923  0.7108 -0.3669 -0.7615 -0.9899  0.1169
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]