Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.9306  0.1518 -0.7552 -1.9402 -0.1299 -2.2783 -1.0086  1.4884 -0.1590 -0.2300
#>  1.4108 -1.2547  0.5665  0.1392  1.2057  1.8418 -0.8005  0.7363  1.9385 -0.4611
#>  0.1110 -0.8230  0.7362  0.2851  0.2604  0.5237 -0.3924  0.6643 -0.9185  0.5677
#> -1.3308 -1.0076  0.5456 -0.7769  1.1899 -0.3692  0.4500 -1.6082 -0.1658  0.8237
#>  1.5514  0.7485 -0.4147 -0.3577 -1.4083 -0.7367  1.0434 -0.3423  0.8495  0.2103
#> -0.1808  0.1198  0.2199 -0.3960 -1.9205  0.2193 -0.2460 -0.5733  0.5559 -0.1786
#>  0.5656  0.2852 -0.2909 -0.6199 -0.4740 -0.9785 -1.2621  1.9750  1.9617 -1.3393
#>  1.0946  0.5160  1.3876  0.1881  0.2278  1.2542  0.8061 -0.9667 -0.4538  0.2228
#>  0.9358 -0.0086  0.6981  0.8912 -0.4137  1.0884  1.2449  1.3284  0.6860  1.4097
#> -0.1520  2.2914 -0.4252  0.5794 -2.4804  0.8588 -0.0989 -1.4681 -0.6996  0.2462
#>  0.0684 -1.2224  0.9225  1.0945  0.8367  0.1256  0.4561  0.9900 -1.0253 -0.4623
#>  2.2739  0.3137  0.6833 -0.2749 -0.4157  0.0713 -0.4660 -0.3876  0.7154  0.0700
#> -0.6350  0.4118 -0.2627 -0.6141  0.3134 -0.1685 -0.1005 -0.0060  0.0343 -0.8094
#>  0.1171 -1.0963 -0.3843  2.6365 -0.5680 -2.2750  0.5567  0.1996 -0.6890 -0.4504
#>  0.5022  0.8903 -2.0661  0.3597 -1.2288  0.2956 -0.6512  0.2688 -0.5753  0.2166
#> -1.2895 -0.6673 -0.8367 -2.0597  0.6116 -0.3747  2.4745  0.1773 -0.7064  0.0644
#>  0.9749 -1.4079  0.7478 -0.4117 -1.6842 -0.6756  0.1647  0.3085  0.6058  0.6035
#>  0.8521  0.1314 -1.7171 -0.0194 -0.0220  0.0997 -0.1000  0.4265 -0.8820  1.3296
#>  1.6657 -1.1874 -0.4278 -0.1114  1.7567  1.9931  1.1463 -0.1341 -0.6435  0.8294
#> -0.5445 -0.3211 -0.0764 -0.2298 -1.3397  0.7884 -1.5188  0.6401 -0.0727  0.6893
#> -1.0512  0.0703 -0.7898  0.2678  0.8012  0.2615  0.2666 -0.3954  0.2306 -0.2511
#> -0.3320 -0.1255 -1.7495 -0.4172  0.4725 -0.6433  0.1472 -2.3778  1.1339  1.7472
#>  0.8431 -0.1426  0.3521 -0.7094  0.2996 -0.5000  0.6806  0.2611 -0.5903 -1.2250
#> -2.4945  0.2061  0.5904 -0.6341  0.8098 -0.3276 -2.2059  1.7537 -0.8951 -1.0096
#> -1.1467  0.5717  0.2228 -0.8058 -0.3178 -1.1903  0.0411 -0.0421  0.6824  0.8572
#> -0.3903  0.5987  0.5516  0.2663  0.5589  0.3115 -0.4187  2.3480  0.0280  0.4594
#> -0.5479  0.4580  0.9728 -0.3163 -3.1807 -0.5366 -1.3625 -0.1099 -0.0586 -0.8695
#>  1.9111  2.6022 -1.9444 -2.0111 -0.1676 -1.3946 -0.4851  0.7692 -0.1320  0.2928
#> -0.0060  0.3717 -1.2258  1.1927  3.0017 -2.2671 -0.7826 -0.1894  0.1629  0.1838
#>  0.9818 -0.6736  1.6460  0.1788 -1.1953  0.3840  0.4182  1.2119  0.7865 -0.5472
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]