Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 5.8553e-01 -1.2629e+00  3.9954e-01  2.1177e+00  6.9697e-01 -1.3812e+00
#>  2.2998e-01  2.5684e-01  2.0178e+00  1.5984e+00 -2.8167e-01  3.6689e-02
#> -4.2420e-02  1.1384e+00  1.5661e+00 -4.5719e-01 -1.0192e+00 -2.1395e-01
#> -1.0164e+00 -4.9367e-01 -4.2933e-01 -8.9357e-01  1.2617e+00  7.8998e-01
#>  4.3140e-01 -8.6520e-01  5.9451e-04 -1.1422e-01 -6.0801e-01 -8.4700e-01
#>  6.9478e-02  5.7202e-01 -5.1293e-02 -5.4139e-02  9.8810e-01  4.0453e-01
#>  7.2500e-02  8.2922e-01 -1.7323e-01  4.1590e-01  3.9037e-01  8.8368e-01
#>  8.6510e-02  5.1604e-01 -1.0567e+00 -1.8587e+00 -1.4291e-01 -1.7740e-01
#> -1.2110e-01 -1.0324e+00 -1.5772e-01 -1.0076e-01  1.3763e+00 -7.9726e-02
#> -9.5454e-01 -1.1228e+00 -5.7907e-02 -1.3209e+00  5.4914e-01  7.8557e-01
#>  8.0433e-01  3.3872e-01  2.5276e-01  7.2916e-01 -1.4368e+00 -6.1130e-01
#> -1.3811e-01  6.9442e-01 -1.9290e+00 -1.2844e+00 -1.0074e+00  1.9130e+00
#>  1.6975e-01 -7.1954e-01 -6.5972e-01 -9.3116e-01 -9.8500e-01 -9.7015e-01
#>  1.0693e+00  6.9709e-01  4.6962e-01 -5.0501e-01 -1.7654e+00  5.4440e-01
#>  5.8152e-01 -9.5202e-01 -2.5223e-01 -6.8154e-01 -8.2571e-01 -1.1008e+00
#> -1.0299e+00 -2.1232e-01 -8.0739e-01  1.1931e+00 -6.3213e-02 -1.9522e+00
#> -1.8820e-01 -8.4123e-01 -3.1141e-01 -1.3870e+00  1.9536e+00 -4.0848e-02
#>  1.6864e+00 -4.2186e-02 -5.9590e-01  2.2519e+00  3.3099e-01  6.9891e-02
#>  1.4226e+00  6.6717e-01  5.8327e-01  2.0133e+00 -5.8678e-01  1.4417e+00
#>  4.5245e-01  5.6962e-02 -1.3214e+00  7.5261e-01  1.2403e+00 -6.7547e-01
#> -3.3629e-01 -5.4379e-01 -1.6173e+00  5.9632e-01  6.6533e-01  1.7406e-02
#> -2.2240e-02 -2.5338e-01 -9.7849e-01 -9.6730e-01  1.7388e+00  1.6689e+00
#> -1.7980e+00 -1.3771e-01  2.0941e+00 -7.9294e-01  3.5922e-01 -7.4048e-01
#> -4.3332e-01 -2.1457e-01  1.4685e+00  1.6755e+00 -1.0806e+00  2.9493e-01
#> -1.9486e+00 -1.3884e+00 -1.5947e+00 -2.0465e-01  4.2155e-01  3.1210e-01
#> -1.8818e+00 -4.5058e-02 -7.6469e-01 -1.0375e+00 -6.7531e-01 -5.1293e-01
#>  5.2472e-01 -9.9817e-01 -3.8512e-01  1.7788e+00 -3.5420e-01 -1.1660e+00
#>  1.2022e+00 -2.5399e-01  4.3968e-01  8.6466e-01  9.1217e-01  1.2281e+00
#> -9.1364e-01 -5.9534e-02 -1.6058e+00 -2.9230e-01 -9.3871e-01  5.3371e-01
#>  1.0659e+00 -1.0297e+00  1.0371e+00 -5.6474e-01  5.9114e-01 -6.2419e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]