Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2178 -1.3714 -0.2030  1.1416 -1.1881  0.6356  0.1943  0.5249 -0.6097  1.6334
#> -0.5242 -0.8762 -0.0805 -0.3700 -0.3524 -2.3987  1.7179  0.6941 -0.1340  0.2981
#>  0.7336 -1.2127  0.7880 -0.5877  0.0160 -0.9451 -0.6341 -0.4170 -1.0967 -0.0172
#> -0.1217 -0.3129 -2.5238  1.1126 -0.8558  0.7863  0.6366  1.9761  0.0890  0.0330
#>  1.0005 -0.7046  0.4249  1.4023  0.1921  0.4870  2.9016  0.6108 -0.1157 -0.2125
#>  0.3103  0.3383  0.2442 -0.4153  0.3088 -0.0529 -1.2234 -0.2120 -1.2273  1.1889
#>  0.6778 -1.5098 -0.1661 -1.7181 -2.3266  1.9362 -1.2691 -0.2016 -1.0805 -0.1096
#> -1.1129  0.8120  1.6013 -2.1480 -0.1308  0.8036 -0.6103  1.4835 -0.0339 -0.4185
#>  0.5730 -0.3788 -0.7575  1.3047  0.7294 -0.5943  1.3612  0.1343  0.3782 -0.2522
#>  0.7864  1.7536  0.9724 -0.2469 -0.7170 -0.1128 -1.1042 -1.0743  0.2194  0.0585
#> -0.5047 -0.6115 -2.4501 -0.1990 -0.8529  0.3462 -0.0114 -0.5170  0.7918  0.3366
#>  0.5471  1.9588 -1.5048  1.5231 -0.1139  2.1569  0.2538 -1.7914  0.0290  1.1988
#> -0.4714  0.8023  0.7323  0.2975  0.5577 -0.0987  1.3407  0.1129 -0.2398  0.7322
#> -0.6559  0.1159  0.4094  0.2942 -0.8048  1.6276  0.0171  0.7562 -1.7830  0.2192
#> -0.7214  0.9354 -0.5315 -0.1462  0.9888 -2.1616 -1.5212 -0.8468 -1.2470  0.3461
#> -0.0514 -0.3294  1.2106 -0.8448 -1.1145  0.2338  0.4808  1.4625  0.3744  1.6258
#>  0.7523  1.0655  1.8394 -1.5839 -0.8591 -1.3027 -0.2693 -0.0297 -0.9723 -0.8686
#> -0.2727  2.3038  0.0965  0.9645  0.7880 -0.9287 -1.3636  0.9066 -0.4961  1.1699
#> -0.9680  1.9549  0.5867  1.6673 -0.6046 -0.0646 -0.1339 -1.9684 -1.4855  1.2029
#> -0.0921  0.4677 -0.1112  0.0540  0.1371 -0.0962  0.9408  0.4414  0.6511  0.0560
#>  0.1789 -1.4773  1.1711 -0.5554  0.4838 -1.1367 -0.1580 -0.3925  0.4007  0.8594
#>  0.2979 -1.0255 -0.4559 -1.3648 -0.2849  1.2064 -0.9190  1.3562 -1.0321  0.3290
#>  0.3616  1.7639 -0.2002 -1.9759  0.7273 -0.0529 -0.6518 -0.9163  1.9450 -0.6158
#>  1.2524 -0.5658  1.8059 -0.2356 -2.2197  0.8259 -1.0412 -0.3597  0.2189 -0.7468
#> -0.8976  1.9348 -0.4728  2.2762  0.1161 -0.3344 -0.4299 -0.5037 -1.3123 -0.0871
#> -0.0865 -1.5511  1.4350  0.0207 -1.2400  1.2293  0.8026 -0.8781 -0.5219  0.1938
#>  1.8968  0.3798  0.0183 -0.5721 -1.0879  1.6152 -1.3478  0.1835  0.6652 -0.4056
#> -0.5884  1.5207  0.5405  0.9837 -0.1259  0.3150 -0.9869 -1.2280 -0.5434 -0.6188
#>  0.4682  0.4488 -0.2726 -0.6207 -1.3311 -0.6211 -0.1919  1.4084 -1.2642  0.9824
#> -0.5434 -0.6009  0.2851 -1.5448  1.1004 -0.6765  0.2708 -0.5582 -1.2895 -0.3082
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]