Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.2403  0.1875  0.6746  0.9158  0.4982 -0.9632  0.1374 -0.0056 -0.3667 -0.7292
#>  1.7923  1.0799  0.2683  0.4906 -0.8525  0.1267  1.1860 -0.2716  1.6313  1.1659
#>  1.2021 -0.9452  1.6910  0.0232 -1.3992 -1.4396 -1.1791 -0.6347 -2.0571 -0.3429
#> -0.4987  0.6215  0.4847 -0.2052 -0.0739 -0.6447 -0.7918 -0.4121  0.7434 -0.3083
#>  0.1323 -0.7818 -0.9464 -1.4441  0.0733  0.2830  0.8692 -0.0237 -0.0955 -1.2752
#>  0.6859 -1.7782 -0.7687 -0.1824  1.4489 -0.1466 -1.3134 -0.8578 -1.1444 -0.7938
#>  1.5575 -0.2451  1.3907 -0.8308  0.2980  0.2310  0.2850  1.0578  0.7308  0.2606
#>  1.7396  0.2587 -0.4692 -1.1709 -0.1020 -0.7587  0.1374 -0.9421  0.5626  1.3039
#>  0.7577  0.1428  1.1280 -0.1735 -0.4493 -1.3806  0.6135  0.2153 -1.2688  1.7539
#>  2.2074 -1.3139 -0.4057 -1.4051  0.2480  0.3860 -0.5884 -2.0791  1.2934  0.3179
#>  0.9592  0.1219 -0.5247 -0.4860 -0.3631 -0.0232  0.3243  0.3956  2.0390  1.0097
#> -1.6221 -0.3689 -1.2912 -0.5745  2.0363  0.5243  1.1603  1.4096  0.2181  0.7160
#>  1.4116 -0.3384  0.0248  0.0208 -0.4511 -0.1098  2.0209 -0.9123  0.3780 -1.3554
#> -0.7673  0.6134 -0.7549  1.4863  1.7769 -0.1772 -1.1539  0.5290 -0.4477 -0.1434
#>  0.0786 -1.2150  0.2545  0.2092 -0.8705 -0.1632  1.1576  0.4218 -1.9859  0.0167
#> -0.7309 -0.8816 -0.3644  0.9358 -0.9036  0.4178 -1.3635  0.4861 -1.0088 -0.5641
#> -1.3427  1.1134  0.4371  0.2449 -0.6227 -0.0868 -0.2213  0.9144 -0.7001  0.0891
#> -0.9816 -0.8730 -0.2503 -1.2360  1.0064  1.3921  2.4869 -0.8043 -2.9608  0.6417
#> -1.5796  1.3082 -0.4455 -1.1600  0.8398 -0.3808  0.7011 -0.0731 -0.7547  0.0898
#> -0.0733  2.0720 -1.5811 -0.9163 -0.7686  0.0321 -0.0280 -0.0423 -0.8364  0.0395
#> -1.7080  0.5968  0.8155  1.8046 -0.4374  0.5159 -1.4819 -0.0771  1.1042 -0.8673
#>  1.0157 -0.0767  0.8438  0.1640 -0.4076  2.3456 -0.4294 -0.0865  0.1207  0.1499
#> -1.4111 -0.0936 -0.8697  1.1448  0.4789  0.5653  0.5816  0.6542  0.1529  0.9593
#>  0.8672 -0.1788 -1.3074 -0.0080  1.2018 -0.3519  0.5554  0.0966  0.3767  0.9471
#>  1.9533  1.5929  0.8588  0.9118 -1.0891 -1.4956 -0.5257  0.3786  1.2602  2.3186
#> -0.7892  0.0316  0.0768  0.5970 -0.1782 -0.2517 -0.2847 -0.0761 -2.4484  0.9639
#> -1.2518  1.0806  2.0955  0.3745 -0.4199  0.7161 -0.6978 -0.7119 -0.0843  0.4504
#> -0.4555 -0.6112  0.2099  1.6932 -1.3169  1.1922 -1.2351  0.1945 -1.0275 -0.8482
#> -0.6440 -0.7016  0.3570 -2.3694  0.5846 -0.3763  0.8922  0.1694 -0.3332 -2.1098
#>  0.7656 -0.0148  1.0551  0.7073  0.4318 -0.5487  1.5614 -1.4700 -0.4855  1.0523
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]