Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6 6.1329e-02  1.3673e+00  1.1612e-02  1.6107e+00  2.1680e-01 -2.9982e-01
#>  9.7091e-01 -7.8615e-01 -1.6593e+00  3.3089e-01 -1.1663e+00 -3.6394e-01
#> -1.2078e+00  1.4064e+00  2.3188e+00  8.7563e-01  4.3208e-01  8.4361e-01
#>  6.8117e-01 -3.4930e-02 -2.3113e-01 -8.3664e-01 -1.8101e+00 -1.0955e-01
#>  5.6990e-01  3.6704e-01 -6.3374e-01  9.8319e-01 -3.5846e-01  1.0586e+00
#> -1.3225e+00  2.9428e-01 -6.4883e-01 -9.8986e-01  2.1696e-01  7.1946e-01
#> -1.2327e+00 -9.7322e-02 -1.9079e+00 -1.1282e+00  1.1718e+00 -6.4271e-01
#>  4.5130e-01 -7.6566e-01  3.6957e-01  3.7738e-01  9.6139e-01  7.7030e-01
#>  4.2225e-01  1.9700e-01  5.8244e-01  2.1742e+00 -2.1612e-01  5.5888e-01
#> -4.2831e-02  1.1592e-01 -3.3520e-01 -1.6493e+00 -1.1304e+00  2.7268e+00
#> -5.3649e-01  5.8785e-01  7.3629e-01  1.1387e+00  4.3568e-01 -1.6709e+00
#>  9.1851e-01 -1.4932e+00  2.1123e+00  1.1338e+00 -3.7431e-01  1.2126e-01
#>  6.0097e-01  1.1974e-01  8.2852e-01  1.4771e-01  2.3607e+00  5.6682e-01
#> -1.5608e+00  1.3432e-01 -1.1205e-02 -2.0693e+00  6.1431e-01 -5.1183e-01
#>  1.4935e+00 -7.0164e-02  1.5647e-01  4.5798e-01 -1.3600e+00  3.1900e-01
#>  2.5402e+00 -1.6224e-01 -3.9686e-01 -1.3377e-01  5.4919e-01  2.8008e-01
#> -1.1340e+00 -3.6241e-01  4.2486e-01  1.3967e-02  1.2203e-01 -1.8697e+00
#>  4.9275e-01 -2.5927e-01 -1.9795e+00  7.9229e-01  5.9343e-01  1.6816e-01
#> -1.4121e-01  9.2154e-02  1.0562e+00 -5.3607e-01  6.0247e-01  9.7103e-01
#>  5.6908e-01  4.4490e-01  6.1546e-01  7.4250e-01 -3.1689e-01  1.2911e+00
#>  2.0097e+00 -6.9749e-01  5.7125e-01  8.2729e-01 -8.4588e-01  2.4921e+00
#>  7.2022e-01  8.1655e-01 -1.3862e+00 -1.8273e-01  4.7668e-01  9.1727e-01
#>  6.6937e-01  9.8115e-01 -6.9881e-01  3.0213e-01  1.8578e+00  7.2619e-01
#> -1.2408e-01  7.4883e-01  5.2812e-01  1.2445e+00 -1.9383e+00 -2.1012e-01
#> -4.7602e-01  5.9548e-01 -5.6912e-01  9.0414e-01 -1.2633e+00 -7.7543e-01
#> -3.8172e-02 -6.5663e-01  5.4071e-01 -1.3305e+00  6.2284e-03 -1.9386e-01
#>  2.4202e-01  2.7702e+00 -2.9632e-01  1.8980e+00  6.7462e-01  2.9467e-01
#>  4.2234e-01  1.0994e+00  3.0024e+00 -7.3371e-01  1.0488e-01 -7.7450e-03
#> -1.4512e+00  1.3690e+00  2.4543e-01 -6.4777e-01  3.9814e-02  1.5155e+00
#> -1.3157e+00 -1.7261e-01  5.5103e-01  9.6407e-02  5.9353e-01  1.6534e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]