Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-2.1726 -0.9104 -0.8498  0.7856 -0.9584 -0.1029 -0.5890 -0.2067 -1.2636  0.3162
#>  0.4392  0.2179  1.4850  0.9034  0.0134  0.7284 -0.6926 -0.8392  0.1337  0.4671
#> -0.2542  0.4720 -0.3633  1.6828 -1.2642  0.5797 -1.3220  0.6449 -0.0754 -0.3356
#> -1.3792 -0.7727 -0.7582 -2.2930 -1.8197  3.1616 -0.1566 -1.8868  2.5209  0.2853
#> -0.0920  0.7427 -0.7095  0.1643  2.1450  0.6183 -1.2834 -1.5662 -0.9647  1.5955
#>  1.0074  1.1750 -1.2493 -0.6507 -0.8068 -0.9692 -1.0621  2.3522 -0.1869  0.6704
#>  0.4733 -1.3934 -0.1248  1.9525  0.3413  0.7372  0.4403  0.9880 -0.7051  1.2830
#> -0.4923  2.5870  1.9809 -0.0897  0.3735 -0.1753 -2.0096  0.5417  1.3061 -2.6595
#> -0.9781 -0.3679  0.0944 -0.1461  0.4757 -0.3779  0.5932 -0.7620 -0.4798  1.8724
#> -1.3609 -0.9224  0.1275  0.0456  1.6748  0.0557  3.0707 -0.8720  0.6062  0.6363
#> -0.2902  0.7014 -1.9758 -0.5429  0.7821 -1.0614 -0.7107 -0.2023  0.7293 -2.2116
#>  0.9231  1.4099  0.4960 -1.5718  0.1856 -1.0373 -0.7599  0.8552  1.0773 -0.5030
#>  2.5269  0.2255 -2.4225  0.9356 -0.1365  0.0974  1.3193 -0.2378 -0.7437 -0.6552
#> -0.1462  0.4985 -1.4086 -0.9227 -0.2122 -1.2798 -1.7347 -1.9531 -0.1449 -0.2003
#>  1.6088  0.3803 -0.5883  0.2968  0.1544 -0.5605 -0.8411 -1.4791  0.7576  0.2014
#>  2.9581  0.7688 -2.2459 -0.3597 -1.3255 -1.1252  0.1324 -1.4499 -0.3330  1.2129
#>  0.5570 -1.6347  2.0686 -2.0454  1.6707  1.1089  1.1442  0.3535  0.1777  0.1966
#> -0.3694  0.6953 -1.5432  0.7512 -0.4617 -0.7588  0.7761  0.2000  1.1145  0.4712
#> -1.0157  0.8737  0.1437 -0.3798  2.2281 -1.2815  0.4004  1.4321 -0.2905  0.3396
#> -0.1021 -0.7978  0.0570  1.0231  0.5720  0.0464  0.4713  0.7827  0.2031 -2.0665
#> -2.4535 -2.3498 -1.5687 -0.6867 -0.3510  0.3805 -0.1025  0.6724 -0.8229  0.0025
#> -1.0174  0.0494 -0.0518 -0.5782 -0.1027 -0.4654  0.3051  0.5847 -0.9939  0.6449
#>  0.6411  0.9782  1.9267 -1.0057 -1.3378  0.0139  1.1775  0.0064 -0.5941 -1.2537
#>  0.8541  0.4403 -0.3101  1.7504  0.4684  1.0871  0.0475  0.8795 -0.1129  0.9587
#> -0.0795  0.7603  0.4405  0.7021  1.4374 -1.4913 -1.7578  0.8413 -2.3742 -0.9568
#> -0.0653 -1.4485 -1.6340 -0.4207  0.4408 -0.8699  0.0332 -0.3040  0.1403  0.9050
#> -0.5890 -0.8001 -0.1388  1.7836  0.5761 -0.5135 -0.5593  0.5003  0.1482  1.4958
#>  0.6783  0.5770 -0.1847  0.9988  0.7617 -0.8870 -0.4284 -1.9800  0.1625 -0.4927
#> -0.1104  0.0227  0.3561  1.3382  0.2809 -0.3929  0.5589  0.2096 -0.9771 -0.8658
#> -0.5836 -0.6293 -0.5748  0.8388 -0.0007 -1.1937  1.8493  0.2373 -0.9224  0.6759
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]