Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.8063  0.6870 -0.2101  1.3234 -0.5440 -0.7171  0.6765  0.0242 -1.2723 -0.0887
#>  1.4142  0.1932  0.2875  0.0491 -0.4638 -0.2346 -0.6751  1.1251  0.5846 -0.9567
#> -0.6617 -0.8544  0.5890  1.6970 -1.2260  1.3105  0.9831 -0.9773  0.6218 -0.9425
#>  0.3675  0.4490 -0.6133  0.3541 -0.3239 -0.2341  0.6038  0.1906 -0.7043  0.1297
#>  0.3275  1.9210  1.2363 -1.0524 -1.0536  0.6966 -0.9603  0.5067 -0.5457 -1.1390
#>  0.9299 -0.2248 -0.6954  0.0566 -0.4849 -0.9553 -0.2268 -0.5096 -0.3005 -0.9739
#>  2.4261  0.6406 -1.0259 -0.3709 -0.4296  2.5647 -0.5056 -0.3297  1.2513  0.6006
#> -1.5294 -1.2099 -1.7617 -0.8528  0.5929  0.2222 -0.4256  0.3601 -1.5127  1.9773
#>  0.2000 -0.5528  1.2968  1.1832  0.1610 -0.0609  0.0988 -1.5723 -1.6318 -0.6936
#>  0.4112  0.8127  0.0844 -0.2746 -0.3314 -0.6686  0.7428  0.2501 -0.2416  0.2723
#> -1.3951  0.7027  0.3883 -1.2438 -0.3473 -1.2625  0.2172  0.7418  1.4468  0.8976
#> -0.5544 -3.2495 -0.9423  0.6344  0.4141 -0.4955 -0.2367 -1.5715  0.3275 -0.3937
#>  0.4160 -0.2953  1.1283  0.5199 -1.5081 -0.6126  0.3727 -1.4852  0.7333  0.1811
#>  0.3206 -0.4307 -1.4900  0.9016  0.1073 -0.0971  0.8264  0.2558  0.9925 -0.4308
#>  0.5593 -0.7755 -0.6166  1.6725  0.8672 -0.2854 -1.0782  1.6868  0.1875  1.1462
#>  1.2110  0.0428 -0.7910 -1.1595 -0.8510  0.1728 -1.5217  0.4523  0.9286  0.4087
#> -0.1459 -0.5536  1.1188  0.6063 -1.0159  1.9386  0.3328 -1.4988 -0.1523 -0.6102
#>  0.0804 -0.7896  0.9107  0.3899  2.0102  0.8928 -1.0183 -0.6428  0.4166  0.9511
#>  1.0162 -1.5138  0.9642 -0.3698  0.0192 -0.7474 -0.1300 -0.4212  0.7238 -0.5597
#>  1.0595 -0.5710  0.1093  0.9822  0.3248 -1.0077  0.9508 -0.1185 -0.2410 -1.9374
#> -0.3146  1.9655 -1.5614 -0.0729 -0.3033  0.0220  1.6826  1.2480  0.2672  1.5440
#> -0.6213 -1.8787 -0.7797  0.1660  0.1801 -0.3941  0.2867  1.1373 -0.7494  0.2859
#> -0.0452  0.8581 -0.6975  0.2609  0.8572  0.5071 -0.4390 -0.3971 -0.2948 -1.4888
#>  0.6136 -0.1410 -0.1767 -1.2436  0.9928 -0.7547 -0.7078 -1.0279 -0.4639  0.6009
#> -0.7318  0.3521  1.1834  0.3520 -1.0288  0.9808  2.2880  1.7219 -1.5487  1.3345
#> -2.0798  0.0847  0.1945 -0.6153 -0.3573 -0.3615  0.5455  0.0376 -2.1483 -0.1216
#>  1.7004 -0.2082  0.2635 -1.8343  0.6226 -0.9840 -2.8282  0.2447  0.8586 -0.3720
#>  0.2387 -0.6440  0.6566 -1.2219  0.8102 -0.6130  0.4175  1.5691  1.8212  0.2517
#>  0.3301 -0.5458  1.3968 -0.8354 -0.0643 -1.7652 -0.0180 -0.8783  0.2210  0.3522
#>  1.7624  0.3073  0.3018  0.0909 -1.7458  0.5494 -0.6088 -1.8967 -0.1576  0.7989
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]