Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.8331  0.0623 -1.3873  0.0126  0.8528 -1.8805  0.7829  2.7225  0.2950 -0.3453
#>  1.7727  2.6032 -1.4636  1.8276  0.7890  0.3102 -0.2202  0.5339  0.8012  0.1782
#>  0.3311 -0.4775  0.7938  1.3378 -1.7127  0.1410  0.9999 -0.7853 -0.6608 -0.8017
#> -1.0640 -1.0579  0.0362 -0.6511  1.5410  1.4091 -0.7256  0.0232  1.5421 -0.3284
#>  1.6299  0.8862  1.2419  0.4210 -2.2636 -0.8835 -0.7461  1.6862 -0.6616  1.6082
#> -0.2301  0.5984 -0.8468  1.9564  0.0910 -0.5704 -0.1351 -0.7051  1.9288 -0.0523
#>  1.2730 -1.3599  0.5294  0.6430 -0.7138 -0.1706  0.9918  1.4224 -0.0638 -0.4089
#> -2.7658 -1.1409 -0.3253  0.5376  0.5723  0.1126 -0.3308 -0.0315 -0.8279  1.1905
#> -0.6036 -2.6988 -0.8816 -0.7194 -0.9023  0.7559  0.4964  1.2143  1.6505  0.7939
#> -1.1744 -1.2137 -0.2990 -0.4166  0.9479  0.6028  1.2009 -0.0772 -0.6226 -0.1971
#>  0.3967 -1.0570 -0.6499  0.1672 -0.7467 -0.1451 -0.4503  1.9234  0.2360 -1.4276
#>  0.2999  0.8521  0.0241  0.0528 -1.3574 -0.3363  0.3635 -0.2362  0.3237 -1.1963
#>  1.5697 -1.8404 -0.7348  0.5232 -0.8890 -1.5771 -0.8115 -0.3306 -0.2040  2.1712
#> -0.3944 -0.2337  1.0420 -0.5191 -1.5935  0.0081 -0.1371  1.1078 -0.8838 -0.8180
#>  0.2384 -0.3649 -0.6344  0.2811  0.1061  0.3599  1.0001  2.6734 -0.7995  0.3798
#> -0.2729  1.0448  1.5325  0.7141  0.0763  0.3695 -0.5217  1.8028  0.0900 -0.1546
#> -0.2577 -0.2979 -2.3553 -1.0526 -1.5441 -0.2320 -1.3460  0.6953  0.5993  0.2429
#>  0.0718  1.6944 -0.2526  0.4082 -0.8674  0.3097 -0.7648 -0.7733 -0.6004 -0.6070
#>  0.0897  1.2383 -0.7775 -0.4114  0.4283 -0.5051  1.0745  1.5654 -0.3033  0.8311
#>  0.5770  0.8645  0.9467  0.1444 -0.1896 -0.6065  0.3473 -0.3299  0.4852  0.5416
#> -0.2252  1.8230 -1.8057 -1.2552 -1.5256 -0.4444 -0.4572 -0.1734  0.0285  1.0417
#> -0.9796  0.5731  0.7618 -0.6611  0.0422 -1.1972  0.9817 -1.3474 -0.9852 -0.1306
#> -0.2192 -1.2250  0.2198  0.2333 -0.4096  0.1819  0.3538 -0.9778  1.9461  0.2227
#>  0.9338  1.3398 -1.0984 -1.1509  1.9911 -0.2545  0.0082  1.6277 -1.7109 -0.9353
#> -0.3376  0.0979 -1.1394  0.7197 -0.2288  0.1161 -0.8521 -1.8470 -0.1480 -1.3225
#>  1.2133  1.8114 -1.9527 -0.4357  0.7838 -1.2277 -3.1921  0.6217  1.1588 -0.0617
#> -1.2042  1.3814  1.4540  0.0468 -0.8563  1.1041 -1.6429 -0.8007  0.7213  0.5867
#>  0.6051  0.6026 -2.2919  0.1617  0.6453  1.0677  0.4116 -1.4049 -0.0447  1.0115
#> -1.1210  0.2017 -0.7364 -0.2898 -1.2423 -0.1139 -0.3962 -0.7121  0.7363 -0.0187
#>  0.9600 -0.1153 -0.7628 -0.6090 -0.1277 -0.2600  0.4311  1.2556  0.7272  0.8797
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]