Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0640 -1.2344 -0.7799 -1.6723  0.7924 -0.3764 -0.2123  1.8216  0.5283 -1.9633
#>  1.5250  1.0403 -0.0826 -1.4624 -0.6381  0.0143  0.2400  1.5945 -0.0191  0.2729
#>  0.6530  0.3795 -0.3426  0.3305  2.4902  0.3484 -0.1177  0.0254  1.0275  0.5766
#> -0.3869  0.0059 -0.3797  1.0325 -0.3295  1.2457  0.8963 -1.6359  1.0420 -1.3217
#> -0.5236  1.8079  2.3723 -0.0441  1.9839  1.0766 -2.2911  1.6769 -0.4937  0.0023
#>  1.1401  0.4892  2.4998  0.1732  1.1028 -1.8012 -0.7394  0.5946  2.0470 -0.6271
#>  0.7633  0.2846 -1.4994  0.1941  0.0095  0.7857  0.3277 -0.0048 -1.3640  0.4712
#>  1.6261 -0.0043 -0.4366  0.0274 -0.3230 -1.0108 -1.2441 -1.9194 -0.1980 -1.2967
#> -0.3896 -1.3615  1.3592  1.6107 -0.6865 -0.4646  0.0663 -0.8382 -0.1299  0.3225
#> -2.8099  0.7736  0.3728 -0.5325  0.0103 -0.7034 -1.3209  0.1966 -1.6570 -0.1594
#> -1.9328  0.2499  0.5239 -0.5004  2.5881  1.0349 -0.2736  1.0001  0.3635 -0.2396
#> -0.6087  0.1358 -1.0840 -0.8525 -0.1490  0.6851  0.0601 -0.1589 -0.7556 -0.5297
#> -1.5051  0.1496  1.3798 -0.5038  0.3642  0.1270  2.5901  3.0145 -0.8536 -1.5369
#>  0.5746 -1.5222  0.2689 -0.3699  1.2308 -2.0125  0.9251 -1.3728  0.7030  1.1794
#>  0.5177 -0.6654  2.3117  0.2382  0.6317  0.6622  1.0387 -0.4796  0.7824 -0.0909
#>  0.1022  0.8304 -0.2984 -1.2819 -0.3239  0.4899 -0.5625  0.1920 -0.0621  1.4888
#>  0.5285 -1.9344  1.4389 -1.2859 -1.1083  0.1836  0.9473 -0.7382 -0.9965 -0.0020
#>  2.1411  0.2891 -0.7475 -1.1315 -0.6261  0.2923 -0.4965 -0.7318  1.0180 -0.8139
#>  1.4241 -1.4076  2.1762 -0.3916 -0.1749 -0.1859 -0.5040 -1.2187 -0.4873  1.0745
#>  0.2634 -0.7406 -0.1364 -0.6762  0.7418 -0.5128 -0.6160 -1.4585 -1.3092 -0.8240
#> -0.4812  0.5923  1.6253 -0.1810  0.9807  0.6199 -0.1629 -0.7549 -0.5701  0.7391
#> -0.8387 -0.3164 -0.3147  0.6090 -0.9760 -0.8302 -0.7788  0.4918  0.6334 -0.4920
#>  1.5705 -0.3871 -0.4567 -1.2618 -0.0012 -2.9263  0.1974  1.7921 -0.4914 -0.4337
#>  1.7346 -0.5788  0.4382 -0.7359 -0.3523 -1.3703 -0.1533 -1.5588  0.2855 -0.9110
#>  0.0878 -0.0222 -0.9011 -0.8744  0.8248 -0.7624  0.8491  0.9054  0.9448 -0.2741
#> -0.2719  0.7532 -0.1219  0.9556 -0.4820  1.1849  0.9837  1.0915 -0.3553  1.1343
#>  0.0674 -0.0578 -0.8795 -1.3178  2.4146 -0.2016 -1.6664  0.9215  0.1942 -0.2298
#>  0.4903 -2.0860  0.6571 -0.1270  1.4098  0.2849  0.8154 -0.1652 -0.4436  0.1696
#>  0.4235  1.0535 -0.5022 -0.7398 -0.1933 -1.4447  2.0944  1.7678 -0.9626  0.6818
#> -0.7010  0.9483 -1.1025  1.3989  1.2264 -0.4608  0.3657 -1.6915 -0.2130 -2.0091
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]