Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.6147 -1.3128  1.9360 -0.0383 -0.1759 -0.1160  0.1254  0.8593 -0.2422  1.0655
#>  2.3985  0.2740 -0.6999  0.0732 -1.3786  0.2580  0.5632  0.2074  0.1544 -0.2404
#>  0.6894  0.5754  0.9633 -0.9071  1.3430  1.5322  0.2094  0.7258  1.1994 -2.0704
#> -1.5245  1.2231  0.8502  1.4830 -1.4798 -1.0297 -1.0159  0.1798  2.1258  0.5136
#>  0.4489  0.1354  1.3072 -1.1049  0.9383 -0.6492  0.7956  1.4626  0.0062  1.1964
#> -1.7586  0.8166 -0.7755  0.2148 -0.2249  0.5705 -3.2497  0.3197  1.1364 -0.2387
#>  0.0653  1.2589 -1.2518  0.3368  0.7938 -0.3883 -0.1210 -0.2004  0.2050 -0.4717
#> -0.3856 -1.7075  1.3554 -1.1746  0.7447  0.0753 -1.3753  0.2555  0.1240  0.7087
#> -0.0028  2.5449 -1.0489 -0.5000 -1.0543 -1.1426  0.9144 -0.7674  0.7595  2.5523
#>  0.4894  0.4678 -2.1877 -0.4528  0.1076 -0.5758 -0.2758 -0.0449 -0.2960 -0.3376
#> -0.2695 -1.4202 -0.7501  0.3151  1.0517  1.2707  0.4860  0.4378 -1.8492  0.7525
#>  1.8291  0.3715  0.2293 -1.3588 -0.0395 -0.3259  0.2551  1.2583 -1.6550 -0.9064
#>  0.4337 -0.6484  0.4129  0.5032  0.0715 -0.6824 -0.4332 -0.5558  0.2447  0.4868
#> -0.1134  0.8730  1.5937 -1.6532  0.0944 -1.3805  0.3300 -2.3240  0.5629  0.3058
#>  1.5928 -1.0409  1.8574 -0.2929 -0.4324 -0.8101 -0.9795 -0.7998  0.3216 -0.3346
#>  0.7374 -0.2493 -0.2624  0.2717 -2.0276  0.4416  0.1178  0.9726 -0.8766  0.5907
#> -1.0562  1.7894 -1.2071 -0.1558 -0.3557 -0.8797 -1.4047  0.7731  0.5002  0.5456
#>  0.1572  0.9087 -0.8059 -1.4212 -1.2289  0.9535  0.0394 -1.0271  0.0461 -0.2172
#>  0.6849  1.1638  0.6695  0.9948  0.1109  0.2577  0.9166 -0.6685  0.0114 -1.2539
#>  0.0014 -0.3161 -0.7896  0.0112  2.1600  0.8067  1.2496 -1.3637  1.0085  0.9990
#>  0.1899 -0.4530  1.8363  0.6433 -0.9100  0.2021 -0.7024  0.1223  0.6063  0.1619
#>  0.3684  0.6529  1.5726  0.6756 -0.2555 -1.8729  0.4211 -0.5935  0.5663  1.7614
#> -2.2583 -0.1281 -0.2519  0.3620 -0.6737  0.5036 -0.3248  0.3702  1.4398 -0.5887
#>  0.4050 -0.2383 -0.3660  1.7740  1.6474  0.5620 -0.0699 -1.7256 -0.8686 -0.7255
#> -1.3096 -0.5371 -0.7105 -0.6017 -0.8372  1.5570  0.8730  1.8974  0.5252 -0.4090
#> -0.6054 -1.3164  1.9590 -0.2821  0.1338 -1.1969  1.1355 -1.0657 -2.7136  1.3211
#>  1.3623 -0.3769 -1.8350 -0.7614  1.2237 -1.3727  0.3213  1.0014  0.1400  0.7804
#> -2.9581 -2.5133  0.4965 -0.1775 -2.3763 -0.3747  0.2121  1.7119  1.2189 -1.0655
#>  0.1094 -2.2601  1.2091  1.0447  1.5658 -0.5105  0.6022 -0.8843 -2.9925 -1.7420
#> -0.8012 -0.4629  0.6268  0.9149 -0.3184 -0.9523 -0.5716  0.2676 -0.2957 -0.4685
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]