Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6-1.7320e+00  1.2202e+00 -3.0198e-02 -9.3147e-01  4.6604e-01  4.1337e-02
#> -1.7592e+00  2.6252e-01 -1.0227e+00  2.5430e-01  4.0052e-01  8.2226e-01
#>  4.4120e-01  1.4893e+00  1.1730e+00 -4.4586e-01 -6.7117e-01 -1.5635e-01
#>  6.4503e-01  4.3861e-01  1.5091e+00  2.1332e-01  5.3440e-02  5.0059e-01
#>  5.9467e-01 -6.4203e-01  2.3855e+00  1.2253e-01 -4.0384e-02  6.5800e-01
#> -1.8826e+00 -1.6826e+00  1.2542e+00 -5.4982e-01  5.5075e-01  5.9748e-02
#>  4.5767e-01 -9.1080e-01 -1.0618e-02 -8.5183e-01 -1.9279e-01  2.4210e+00
#>  5.9954e-02 -3.3609e-01 -2.9698e-01 -2.3293e+00  6.5270e-01 -6.5099e-01
#>  1.2925e+00 -1.7297e+00  1.0531e+00  1.1331e+00 -3.1744e-02  3.3868e-01
#> -5.1163e-01  5.8990e-01  3.3232e-01 -4.3852e-01 -4.1720e-02  3.4099e-01
#> -5.0210e-01  6.2912e-01  9.1898e-01 -1.3754e-01  9.4684e-01 -1.5712e+00
#> -5.5315e-01  2.4680e+00 -9.6840e-01 -9.7757e-01 -5.7049e-02 -1.2116e+00
#>  3.0448e-01 -2.3207e-01 -4.6078e-01 -1.2002e+00 -8.6724e-01 -6.2294e-01
#> -8.5485e-02  1.0989e+00  9.4204e-01 -1.8546e-01 -6.1862e-03 -2.8771e-01
#> -4.1895e-01 -6.7726e-01 -9.5113e-01  5.9532e-01 -6.4911e-01 -7.6313e-01
#>  4.8606e-01 -3.6667e-01 -2.1614e-01 -4.7743e-01 -1.2843e-01  1.6821e-01
#>  3.6501e-01  1.9477e-01 -9.1599e-01  1.1646e+00 -6.5735e-01 -3.1578e-02
#> -1.1202e+00  3.8715e-01 -2.0309e-01  6.9410e-01  6.8376e-01 -7.0685e-01
#> -4.0095e-02  3.5507e-01 -1.7565e-01  2.1115e-01 -2.8574e-01 -1.0480e+00
#> -2.2894e-01  4.3067e-01  1.0819e+00 -6.6083e-01 -1.5806e+00  1.8721e+00
#>  1.2265e+00 -1.9097e+00  4.5246e-02  6.0037e-01  5.7473e-02 -1.3969e+00
#> -4.9507e-03 -4.3065e-01 -9.8017e-01 -8.1317e-01  1.0357e+00 -1.3624e+00
#> -1.5616e+00 -2.3164e+00 -7.9957e-01  9.2404e-01 -1.0490e+00 -5.3435e-01
#>  2.8858e-01  7.4573e-01 -9.0276e-01 -6.5108e-03  6.5137e-01 -5.1698e-01
#> -9.6295e-02  4.5083e-01 -9.1896e-01  5.4516e-01 -1.6419e-01 -1.4999e-01
#>  1.2427e+00 -1.5953e+00 -4.5004e-01 -2.3247e-02  2.1848e+00 -9.0029e-01
#> -1.0199e+00  1.0209e+00  7.8450e-01 -8.1422e-03 -1.9726e-02  4.8950e-01
#>  2.3145e-01  4.7020e-01  1.3167e+00 -6.3235e-01  2.2762e-01 -2.1840e-01
#> -3.3603e-02  2.3858e+00  5.6171e-01 -9.3601e-01  9.0885e-01  1.2811e+00
#>  2.6721e-01 -2.8563e+00  1.8177e+00  4.8050e-01 -1.5845e+00 -1.0574e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]