Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.6617  0.9437  0.5523  3.3456 -1.8702  1.4527  0.3230  0.4986  0.7009  0.0772
#>  0.4427  0.7201  0.9003 -0.3392  0.8652 -1.4672  0.1759 -0.1263  0.0878 -0.0746
#>  0.1298  0.7520 -1.0758 -0.5558  0.8156  0.5110 -0.3822  1.8922  0.8599 -0.2830
#>  0.2960  1.2040  1.1186  0.8519 -1.1762 -0.4282  0.7223  0.8454  0.0788  1.0758
#> -2.6010 -0.8938  0.3824  0.3654  0.3689 -1.2724  0.4023 -0.0434  0.0151 -1.2142
#>  0.8577 -1.6677 -0.3551 -0.7832 -0.2788 -0.6198  0.1305  1.2934  0.8903 -0.9583
#>  1.4741 -1.3091  1.1801 -1.1696 -0.4515  0.1434  0.0045 -0.5853  0.5431 -0.1955
#>  0.9100 -0.1135 -0.3485 -0.5664  1.1524  2.7094 -1.3091  0.9400  0.0582 -0.1671
#>  0.8223  1.6345 -1.1910  0.5343 -0.4136 -0.4017  0.2284  0.0402  1.3006  0.4326
#>  2.9195 -0.0452 -0.3137  0.6168 -0.9150 -1.0416  2.1645  0.6506  0.1165 -0.8754
#> -1.4144  1.9545 -0.2348 -0.1525 -0.8624  0.2078  0.4041 -0.8781 -0.1977  0.4976
#> -1.1131  0.4173 -0.1725  1.5624  1.7691  1.5874 -0.6746 -1.9240 -0.1051 -0.4961
#>  0.3704  0.7023  0.1325 -0.4172 -0.0804  2.4752 -0.2438  0.1115  1.2779  1.1144
#>  1.1953 -0.0076 -0.1798  0.6575  0.5868 -1.1679  0.6670  0.4660 -1.5770  0.8970
#> -0.2302 -0.5356  0.3361 -1.3185 -0.2037 -0.8813 -0.1862 -1.2390  0.2870 -0.4365
#> -0.6659  0.1125 -0.4104 -2.8284 -0.4696 -1.2868 -0.7512 -1.3240  1.2672  0.4484
#> -0.7195  0.7434  0.2800  1.9203  1.5932 -1.4126 -0.5443  0.1051  0.2497 -0.0372
#>  0.6782 -1.1485 -0.5514 -0.2696  0.2874  0.3820  0.3530 -0.3843 -0.3314 -2.2675
#>  0.7858  0.1781 -1.3775  0.0322 -2.5675 -0.1713 -0.6265 -0.3501  0.5848  0.0203
#> -0.0585 -1.2471 -0.8280 -0.6527  0.6205 -0.9488  0.4340 -0.8701  0.0269 -0.7439
#> -0.2524  1.1448 -2.2199 -1.8866 -0.8760  0.5749  0.6919 -1.1288 -0.1333  0.6674
#> -0.1784 -1.5222  0.5776 -0.1690  0.2641 -0.2648 -0.4791 -0.4550  1.0240 -0.6570
#>  1.5401 -0.1966 -0.0480  1.1096 -1.5501  0.2418  0.5570 -0.7954 -1.0190  3.7412
#> -0.2251  0.0968 -0.6876  1.8971  0.0414 -0.1874  0.9186  1.2830 -0.8158 -1.3359
#>  1.1836 -0.0110 -0.0622 -0.8703  0.7762 -0.8975 -0.0505 -0.5777  1.0315  0.1639
#> -2.7815  0.4925  0.7867 -0.2638 -1.4310 -0.4167 -1.5751  0.1507  0.1024 -0.3973
#> -1.1403 -0.0695  0.4361  0.5213  0.7275 -1.5300 -0.2334 -2.3006 -1.8001 -1.4369
#> -0.0347 -0.0454 -1.3815  1.5063 -1.1740 -0.7447 -1.7689  0.6994  0.4062 -0.2251
#> -0.3932  0.7946  0.8325 -1.3405 -1.5512  0.1545  0.4274 -0.3572  0.0579 -1.6963
#>  0.4279 -0.2474  1.1319 -0.1845 -2.0113  0.1023 -1.3427  0.4613 -1.8927 -0.0066
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]