Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.3258 -1.3780 -0.4874 -1.6245  0.6393  0.2766 -0.2293 -0.2390 -2.5525  0.8025
#> -0.6188 -0.5063 -0.1526 -0.1928  0.8795 -0.3867 -0.5251 -1.3146  0.4999 -0.7844
#> -0.3653 -0.9971 -0.9685 -0.6164  0.5085  1.7593  0.8984 -0.8894  1.9138  0.4262
#>  0.3119 -1.0403  0.6126  1.4511  0.5630 -0.8004  1.9674 -1.5415  0.8485 -0.9487
#> -0.7138 -1.8851 -0.1880  0.5377 -0.4585  0.4456  0.9452  0.8067 -0.1547 -0.1550
#> -0.4621 -1.0318 -0.1259 -0.3037  1.4781  1.0789 -0.9246  0.4753  1.0382 -0.2278
#> -0.2513  0.2711  0.3022  2.4777  1.0123  1.1840  1.1635  0.8601 -1.7219 -0.6588
#> -0.8431 -0.7422 -0.8447 -0.5605  1.1082  0.0847  1.0465  0.3440  2.1883  2.0321
#> -0.3358  0.1056  0.4701  1.5388  0.4606  0.2245 -0.2025 -1.2812 -0.6898  0.1197
#>  0.4493 -0.2349  0.5276  0.3442  0.8276 -0.9254 -0.4664 -0.5554 -1.5615 -0.6802
#>  0.5775 -0.4189  0.8579 -0.5926  0.1378  1.1135  1.0117  0.2551  1.5185 -0.9819
#>  1.4268 -0.4374  0.1886  0.5780 -1.3108 -0.7598  1.5184 -0.6772  0.2750  0.9908
#> -1.7614  1.3747 -2.3194 -0.6879  0.2094 -0.6900  0.4997  0.5341  0.4410  2.0234
#> -0.0388  0.6143 -0.9505  0.1021 -0.0838 -1.7559 -0.7443 -2.2012 -0.8236 -1.3560
#> -1.2034  0.7965 -0.5026 -1.5357  0.3523  0.4743 -0.7295  0.2811  0.1918  1.7374
#> -0.0289  0.1898  1.6826  1.5117  0.6395  1.7951  0.6422 -0.3358  0.2746 -1.5427
#> -0.8057 -0.1631 -0.1647  0.8111  0.7027  0.8682 -0.1549 -0.4352 -1.0960  0.0987
#> -0.4432 -1.3211 -0.3492  0.1046 -0.7759 -0.0375 -2.1252  0.7418  1.9277 -1.5132
#> -0.0558  1.2624  0.7501  0.2840 -0.5282 -0.4937 -1.2025  0.5837  0.5765  0.6177
#>  1.0808 -0.9563 -1.5375 -0.5896  0.1530  0.1194  0.5120  0.5474 -1.7033  0.3223
#>  0.7526  1.0643  0.8586  0.1028 -1.1859 -0.4351 -0.1293  0.3370  0.3813 -1.2464
#> -0.5923 -1.3884 -0.6106  1.0484  0.4133 -0.5572  0.1184  0.1644  0.2680 -0.6627
#>  0.1686 -1.8043  1.1069 -0.5975  0.2216 -0.3282  0.0414 -1.8330  0.3251 -0.3160
#> -1.0432 -1.4866  2.0611 -0.4460  0.0899 -1.8562  0.2047  0.0417  1.6019  0.6125
#>  0.8297 -1.0004 -1.2154 -0.4117  0.4304  1.0202 -0.1374  0.0766  0.3093  1.0833
#> -1.8129  0.2596 -0.9631 -0.5380 -0.6481 -0.9264 -0.0915  0.7137  1.4173  0.2705
#> -0.5090 -0.6741 -0.5305  0.6116  0.3176 -2.2026  0.3306  0.6653 -1.2221  0.3724
#> -0.3961 -1.1967  1.1401  0.0626  0.9964  1.9751  0.2266 -1.6842 -1.6661 -0.8858
#> -0.2930 -0.3194 -1.3697  2.1605  1.1578  1.9825 -0.7530  0.0588 -0.6663 -0.7675
#> -2.0765  0.0312 -0.1012 -0.0272  1.0560  0.4796  1.3146  0.1722  1.6805 -0.7565
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]