Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.4917  2.0255  0.0233 -0.4262  0.4025  1.5907 -0.5875 -0.2811 -0.3286 -0.0628
#>  0.3284  0.6652  0.2505  0.6137 -2.1131 -1.4177  0.2850  0.2977 -1.1183  1.8257
#> -0.1780 -0.6391  0.9505 -2.0409  0.5157 -1.2393  0.7067 -0.1025 -0.9478  0.3039
#> -0.8874 -1.2595 -1.0751 -0.1096  1.9044  1.3717  0.0813  0.4338 -0.1552  0.1081
#> -1.6811 -1.1270 -0.6070  1.7156  0.0796  0.2977 -0.7093 -0.8921 -0.1526  0.0192
#>  0.4178  0.5579  0.1484 -0.3049  1.3526  1.4058  0.5489  0.6360  0.2025 -0.6698
#> -0.0750 -0.6011  0.0590 -0.8211  0.5662  1.2396 -0.7657 -1.4396 -0.2593  0.4086
#> -1.0492  2.5263 -0.5022  1.3157  0.0781 -0.6575  1.5541 -0.4796 -0.5102 -2.5243
#>  0.0393  0.1861 -0.4695  0.0338 -1.0715 -1.0496 -1.3361  0.1652 -1.2700  0.5043
#> -1.4262  0.4725  0.3870 -0.8406 -0.8299 -0.5193 -0.0925  0.5303  0.4223  0.5008
#> -0.2211  0.2047  1.4970 -0.7567  0.0595 -1.5918 -2.8940  0.0367 -0.7582 -1.6573
#> -1.2825 -1.8491  0.4491 -1.8656  0.5207  0.6344 -0.9481 -0.2003 -0.6136 -0.1331
#>  0.3873 -2.2908  0.2737 -3.4150  0.7839  1.1217  0.2985  1.3648 -1.1897  1.5187
#>  0.5880 -2.1032 -1.2000 -0.5322 -1.2478  0.7038  1.0159  1.4498 -0.3929  0.7500
#>  1.3479  0.1030  0.0191  0.4676  1.2160  0.6117  0.1044 -1.3974  0.1255 -0.5037
#> -0.4613  1.0142 -0.9505 -0.2669  0.5644  0.3260 -2.1150  0.0414  0.4707 -1.5646
#> -0.3973 -1.6374 -0.2713 -0.6891 -1.2657  0.4581  0.5862 -0.6353 -0.1283  0.6197
#>  0.1689  1.1526 -0.5667  0.9594  1.0936 -0.4098  1.1252  0.1591  1.3500 -1.0332
#> -0.0423 -0.1502  0.3097  1.0777  0.4047  0.3528 -1.4831  1.7041 -0.7976  1.4094
#>  0.2982  0.1331  0.8111 -0.6454  0.0124 -0.9443  0.1888 -0.3263  0.3532  1.0425
#>  0.9751 -1.1504 -0.0401 -0.1224 -1.0682  1.2035  0.8737  0.5827  0.5272  1.8750
#>  1.9031 -2.7218  1.0364 -0.1807 -1.3076  0.8816 -1.3145 -1.3448 -1.7060 -0.9430
#> -1.3837  1.9636  0.9649 -0.7951  0.0670 -0.1720 -0.3113  1.0290 -0.4358  1.1370
#> -0.0185 -1.0281  1.8125 -1.4165 -0.9591  0.6443  0.6804  0.6865  0.8612  0.5360
#>  1.3060  0.7010  0.8856 -0.2438  0.9551 -0.5456 -0.7471  1.9489 -1.0849  1.0417
#>  0.5244  0.3371 -0.1115 -1.2343  0.1060 -0.4063  1.7539  0.6114  0.6800  0.0039
#> -0.6791 -0.6555  1.2602 -0.1220  0.4616 -0.6798 -0.7913 -2.4126  2.1765 -0.6577
#> -1.7419  0.1443  0.6316 -0.6046 -0.6298 -0.8544  1.3327  0.4363 -1.1919 -1.6425
#>  1.0799 -0.9268 -0.7830 -1.0700  0.0634 -0.6604 -0.4749 -1.6890 -0.5970  0.1707
#> -1.3376  0.5323  0.4308 -1.3873 -0.1235 -0.0228  0.3321 -0.9780 -1.1351  1.4410
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]