Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.7641  0.5973  1.6578 -0.9252 -0.4300 -0.3962 -2.2363 -0.6175 -1.3804  1.4336
#>  0.6134  0.4031 -0.0699 -0.8300 -1.6070 -1.1428 -0.4941 -1.1894  0.6964  1.8466
#> -0.4181  0.2320 -0.8370  1.2144 -1.2397 -0.2080 -0.9541 -1.1618 -0.2339  0.8297
#> -0.2921  0.7832  1.2764 -0.5144 -0.2869 -0.0554  0.4136  0.2233  0.9646  0.6739
#> -0.3777  2.5620 -0.5033 -0.0008 -1.2311 -0.8135  1.4975  0.2215  0.4860 -0.7529
#>  0.4204 -0.5295  0.0362  0.9809  0.5754 -0.4425 -0.1877  1.9563 -0.1148  0.7734
#>  0.1168  0.5628 -0.8547 -1.0553 -0.5691  0.8842 -1.6245 -0.9682  0.4159  0.5160
#> -0.1480  0.3366  1.5084  1.2183 -0.3023 -0.6098 -0.6009 -0.0051  0.0771  0.1871
#> -1.9566  0.7860 -0.6630  1.0773 -1.1362  0.9465 -0.9373 -0.6578  1.0481 -2.4175
#> -0.1807  1.0197 -0.1880 -0.3201  1.0237 -0.3576  1.5444 -1.5351  1.5185  0.5927
#>  0.5100  1.9202  0.5478  0.1076 -1.8397 -1.5652 -0.8671 -1.7745  0.2538 -0.1012
#> -1.7735 -0.0889  0.2582  0.1529  0.6585  0.3329  1.2617 -0.0709  0.7447 -0.6331
#>  0.5537 -0.1950 -0.4670 -0.3505  0.2872 -0.6537  1.3718 -1.5401  1.2658 -0.6532
#> -2.9579 -0.8177 -0.4337  0.3605  1.1683 -0.1958 -0.7624  0.4978  0.7083 -0.8957
#> -2.5609  0.1788 -0.0735  0.1931 -0.1229  0.1825  0.8444 -0.6737 -0.0432  1.3006
#> -0.3432 -0.8737 -0.7086  1.5414  0.1170 -0.5341 -0.1496  0.7755 -0.0188  0.2332
#>  0.2008  0.0965 -0.2278  1.4954  0.0380 -0.3064  1.8089 -0.4010  0.2953 -0.5772
#>  1.4273 -1.8082  0.4785  0.8811  0.6250 -0.6278  0.4834  1.1135  0.1580  2.0534
#>  0.4624 -0.8114  1.0459 -0.4481  0.5531 -1.1256 -0.0715  0.1179 -0.3788  1.0655
#>  0.8668  0.8091 -0.4195 -0.4534 -0.4953  0.7408  0.7099 -1.2497  1.8307  1.3626
#> -0.6619  0.2317 -0.3232 -1.4352  1.2836 -0.6051 -0.4200 -0.8889 -0.0672  1.0592
#> -0.2570 -0.3570 -0.0203  1.0426  2.4914 -1.4169  1.5382  0.1189 -0.2350 -0.2569
#> -0.5655 -1.2713  0.6396  0.5904 -0.4727  1.2027 -0.3595  0.1411  0.8364  0.3820
#>  1.2178  1.2646 -1.2895  0.5111  0.4165 -1.7762  3.2317 -1.1012 -0.2715 -0.7495
#> -0.3355 -0.5795 -0.3760 -1.8155  0.2632 -0.1222  0.9691  0.6362 -0.5230  0.2247
#> -0.6901 -0.8620  0.0506 -0.5004 -0.0602  0.0452 -0.1987  0.1678  0.0886 -0.3258
#>  0.4358 -0.0396  0.1438  0.7161 -0.5593 -0.6654 -1.5379 -0.9101  0.4417 -0.0010
#> -0.3182 -0.9005 -0.1537 -1.9273  0.5782 -0.0530  2.0345  1.3128 -0.5662 -0.1109
#>  0.3996  1.7620  1.5382  0.9010  2.3167  0.2396 -0.6375  1.5668 -1.0171 -0.1534
#> -0.4092  0.1231 -1.0884  0.3730 -0.2554  1.8107  0.3642 -0.3566  0.1124 -0.9650
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]