Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-0.0716 -1.9482 -0.4930 -0.5462  0.7846  1.1117 -0.2389  0.6354  1.8059 -1.7767
#> -0.0125 -0.9543 -0.8283 -0.8317  0.0367 -0.1970 -0.1725 -0.1028 -0.9767  0.1733
#> -0.9383  0.6644 -1.3412  0.5781  0.0517 -1.5613 -0.5809 -1.2715 -1.5557 -0.3814
#> -1.1906 -1.1262  0.6504 -0.0733 -0.7741 -1.8589 -1.1855 -0.4860 -1.4097 -0.3983
#> -0.9606 -0.2253  1.1710  0.1146  0.4780 -1.6771 -0.2162 -0.5521  1.0858  0.7364
#>  1.5148 -1.3757 -2.0425  1.6550  1.3191 -0.7660  1.0304  0.4785 -0.9999 -0.1799
#>  1.1703 -1.3650 -0.9420  1.6768  0.1776  2.3089 -0.2566  1.2093  1.0390  1.4782
#> -0.1109  0.7998 -0.3220  0.7829 -1.8205  0.0417  0.2044  0.7165 -0.1118  1.7416
#> -1.2019 -0.8192  0.2816  0.2411 -0.6457 -0.0536  1.6191 -1.2421  0.5805 -0.4000
#> -0.9480 -1.5011  0.8687  1.1118  1.0105 -1.1639  1.4931 -0.0444  0.7602  1.0116
#>  0.0414  1.6559  0.7963 -0.0274 -0.6624  0.6427 -0.7071  0.8513  0.2563  0.0027
#>  0.9955 -1.0521  0.0516  1.0422 -0.5790 -0.8774  1.0936  0.3802  0.7849 -1.4127
#> -0.8948  0.4407  0.3748  0.8455  0.3026  0.0682 -0.7536  0.1355 -0.4149 -0.4237
#>  1.2971  1.3804 -0.7196 -0.6426 -0.4704 -0.2163 -0.6640  0.2793 -0.3504  0.1226
#> -0.4815 -1.8199 -0.4796 -0.7749  0.0776 -0.6823  0.8338  1.3896 -0.4065 -0.6570
#> -0.7764  0.6697  0.7099  1.5156  0.0618  0.2926  0.3268  0.7762  0.5530 -0.4707
#> -0.3733  0.0989  1.2598 -1.6744  0.9259  0.5487  1.0910  0.4452  0.2722 -0.1801
#> -0.9871  0.0483 -0.1017  0.4041 -0.6374  0.0043 -0.1556  0.3555  0.3336  0.3465
#>  0.8979 -1.4129 -0.8626 -0.3179 -0.4889  0.2707  0.8502  0.2079  0.1377  0.9311
#>  1.5742 -0.3784  1.3604  1.6213  2.0748 -1.4350 -0.5883 -0.0300 -0.6724  0.6661
#>  1.3737 -0.7440 -1.0903 -0.2871  0.5976 -1.1919  0.8375  1.2852  1.4697 -1.2851
#>  1.1483 -1.6419 -1.0072  1.5147  0.5410  1.0413 -0.8203  0.8009  0.8078 -0.0955
#>  0.7861  2.1785  1.2986 -0.5665 -0.6918  0.7801  0.3659  0.3041  0.6896 -0.1489
#> -0.8327 -0.3362  1.1841  0.0672  0.2022 -1.5714  0.4381  0.6780 -0.7866 -0.0094
#> -1.0708  2.0287  0.3375  0.3623 -0.6219 -1.8442  1.4755 -0.0062 -0.4591 -0.0140
#>  0.8844  1.2921 -1.9928  1.1658 -0.5639 -1.0473  0.5634  0.4767  0.8744 -0.0194
#>  0.0396  2.7369  0.3408 -1.1942 -0.1223 -0.8129  0.9847  0.1669  1.3697 -2.1793
#> -0.4773  0.2511  0.5881 -0.1595  0.2292  1.6300 -1.7618 -0.0384 -0.1955 -0.3134
#> -2.0805  0.0079 -1.8923 -0.3768 -0.2115 -1.8010 -0.6408 -1.1603  1.0628  0.9622
#>  1.1091 -0.8312  0.0374 -1.6044 -1.3151 -2.4777 -0.3155 -0.6166 -0.1271 -1.9362
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]