Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.6061 -0.2115 -0.3278 -0.6565  1.3929 -0.0544 -0.4979  0.4145  1.3877  0.1172
#>  0.2595  1.5191 -0.7070 -0.6296 -1.0786 -1.5697 -1.2751  0.9027  1.3866 -0.2459
#>  0.2010  0.3308 -0.5188 -0.3002  1.4886 -0.5563 -0.5247 -0.4023  0.2755  0.0262
#> -0.1336  0.1065 -0.1723  0.7159 -1.0027 -0.6567  0.5579  1.1172  1.9940  0.5289
#>  1.5772 -0.5989 -0.8644 -1.1230  0.9508 -0.6531  0.7772  1.3382  0.0201  0.1229
#> -0.3452  1.5840  0.0426  0.2207  1.2044 -0.3577  0.9678  1.3770  0.7169  0.7156
#> -0.9988 -0.6924  1.1503 -0.2302  0.2705  0.1741 -1.0697  2.3475  2.5944 -0.3005
#>  1.5196  0.0897  0.4798 -0.1889  1.5999  1.6250 -0.5405  1.5611  0.0068 -1.1950
#>  0.1021  1.5648 -0.8156  1.5852 -0.2301  0.2584  0.0978 -0.3633 -0.2652 -1.4552
#> -0.7452  1.0621 -0.5949  0.7520 -1.8138  0.2797 -1.4424  0.0349 -0.7415 -0.2827
#> -1.1557  1.1899  0.5508 -0.4481 -1.5539  1.2608 -1.1247 -0.8751 -0.6055  0.0340
#> -0.3167 -0.0882 -0.3171  1.3980  1.0678 -0.3146 -0.4261 -2.2233  0.7922 -0.5733
#>  0.5424  1.1211  0.8662 -0.1221 -0.0445  0.8753 -1.1984 -0.7924 -1.2973  0.3223
#> -0.3045  0.6733 -0.2013  0.2687 -0.0938 -1.1833  0.1565 -0.7440 -1.3736 -0.6960
#> -0.7682 -0.3773  1.0303  2.3711  0.3307  1.1705 -1.2853 -0.7909 -0.0658  0.7425
#> -0.2409  0.4163  1.2016  1.8187  0.1030  0.2488 -1.1189 -1.2303  0.5404  0.1386
#> -0.2885 -0.6504  0.6703 -0.6871 -1.4301 -0.0642 -1.7157 -2.1478  1.4734 -1.0911
#> -1.6753 -1.7407 -0.8548 -2.0633 -0.0786  0.2228  0.6304 -0.7988  0.0574  1.8983
#>  1.1212  1.1598 -0.6411 -0.8782 -0.2205 -0.2844 -0.3479  0.6661  2.0923 -0.5422
#>  1.9637 -0.5487 -0.7484 -0.1659  0.5107  0.8007 -0.8117  0.1430  0.7997  0.5752
#>  0.0755  0.9363 -0.2422 -1.2351  0.7160  0.9368 -0.7476  0.8469  0.2467  0.1961
#>  0.1582  0.9814  1.0028  0.5190  0.5418  2.0563  0.3083 -1.7407  0.7501  1.0497
#> -0.3195  1.0969  1.0804 -0.6443  1.0912  1.6852  0.6423  0.8936 -0.9019  0.1466
#>  1.2318 -0.5678  0.2114 -0.6973  0.8379  0.8632  0.6901  0.9635 -1.5248 -0.6645
#> -0.2423 -0.2651  1.7402 -0.2901  0.7541  0.3004 -2.6876 -1.0165 -1.2983 -0.3848
#> -1.4499 -1.4863  1.1310  0.8483  0.3651 -1.7018  0.2327  1.4899 -1.6665 -1.4246
#>  1.4942 -0.2910  0.7423 -1.1416 -0.6207  1.6460 -0.4257 -1.5145 -1.1328 -1.0742
#> -1.5573 -0.7063 -0.1337 -0.7242 -0.0918  1.9400 -0.3717  1.8163 -1.6263  0.4281
#> -0.1075  0.8011 -0.7426  0.1341  1.5959 -0.1725  1.5709  0.9450  0.4942  0.0495
#> -0.4727 -0.5255  1.9052  0.3688 -0.7614  0.1553  0.0755  0.7203  0.2082 -0.1218
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]