Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-0.8414  1.4812 -1.6958  1.2784 -0.7717  0.2557 -0.1306  0.7638  0.1928 -0.3661
#> -0.2023  0.1375  0.2884 -0.1011  0.2223  0.3356 -0.9879 -1.6130 -1.4021 -1.3193
#> -0.7468 -0.0026 -0.1964 -0.8007  0.0120  0.7636 -1.5304  1.2688  1.9526  0.7281
#>  0.4520 -1.6716  1.7594  0.2397  0.3954 -0.4806 -1.5789 -0.5127  2.0030 -0.3585
#>  0.7741 -0.0655  0.2172 -1.2392  0.2273 -0.9887 -0.8974  0.3258 -0.9482 -0.0071
#> -0.6761 -1.4172  1.0279 -2.6947  1.3392 -0.9428 -0.4091 -1.4078 -2.5590  0.8261
#> -0.4908  0.2423 -1.2595  0.7091 -1.9881  0.3273 -0.9756 -1.5043  0.8525  0.2449
#>  1.3519 -0.3133  1.0660 -0.3368 -0.1453  0.5400 -1.5076  0.6498 -1.0287 -2.7366
#>  0.3541 -0.0608 -1.6416  2.0981  1.5970 -0.4780  0.4762 -0.2019  0.2797  0.1565
#>  0.5181 -0.4975 -0.9378 -1.3658  1.2002  2.3075 -1.4269 -0.2306 -0.1263  0.1641
#> -0.4220  0.2131  0.0390  0.3993 -0.5299  1.6696 -1.2309 -0.8752 -1.2527  0.7651
#> -0.3390  1.2737 -1.2785 -0.6194 -0.4878 -1.2801 -0.0058 -0.3281 -1.0427  1.1923
#>  2.5325 -1.4313  0.0389 -1.7831  1.4062 -0.0626  0.9208 -0.5080 -0.1659  1.1013
#> -0.7132  0.2011  0.0051  0.3182 -0.4627 -0.6532 -0.7145 -0.4613  0.4758  0.0917
#> -0.7464  0.0417 -1.4656  0.6014  0.3455 -0.0327 -0.3929 -1.3853 -0.2846  0.8871
#> -0.6793  1.0672 -0.1506  0.1154  0.5958 -0.3262 -2.6401  1.4757  0.7559  0.5862
#> -1.0696 -0.3644  1.7769  0.9397 -0.4379 -1.2269  1.1511  0.9154  0.7638  1.0988
#>  1.0404 -0.0365  0.1682 -2.1945  1.4703  0.6182  0.6352 -0.7549  0.6350 -0.8046
#>  0.5230  1.6516 -0.7968 -0.1608 -0.5987  0.4548 -0.3340  1.2126 -1.4726  0.1838
#> -1.1552  2.3342  0.7766  0.9601 -0.3940 -0.7583  0.6415  1.1800  0.0129 -0.0673
#> -1.0766 -0.6951 -0.6223  0.9910  0.6838  0.1530 -1.6243  0.1837  1.4935  0.3211
#>  1.6339 -0.9965  1.2633  0.7150  0.5704 -1.5190 -0.5542 -0.8039  0.7528  1.2303
#> -1.5290 -0.9175  0.6966  0.3976  1.7433 -1.2090 -1.2260 -0.9857 -1.6348  1.0959
#> -1.1520 -1.2063 -1.2389  0.5149  0.1382 -0.5833  1.2023  1.3273 -0.2536 -1.0071
#>  0.4112 -0.4622 -0.5568  0.3038  1.6954  0.0444 -0.9898  0.2250 -1.1644  0.2562
#>  2.0203  0.4075  0.2889  0.2925 -1.0324  0.1966  0.5593 -0.0832 -1.2322 -0.4708
#>  0.0379 -2.1018  1.0520  0.0778  0.9554  0.1239 -0.0731 -0.5759  1.8954  1.5127
#>  0.7003 -1.8516 -0.7422 -1.6257  0.1449  0.4960 -1.5010 -1.2631  0.0495 -0.3830
#>  0.1347 -0.3641 -1.4580 -2.1396 -0.1062 -2.2750 -0.5127 -1.0081 -0.1652 -1.0369
#>  0.4706  1.1616 -1.2821 -1.8567  0.2367 -0.0311  0.5192  0.3748  1.5510 -0.2714
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]