Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.1598  0.9588  1.1969  0.1343 -0.1364  0.3002  1.1671 -0.1161  0.0765 -0.2103
#> -0.0548  1.2704  0.9514 -0.0963 -1.1346 -0.2585  0.6275  0.3914 -0.0481  0.3050
#> -0.9061  0.4989 -0.4300  1.0237  0.5058  0.3088 -1.3828 -2.9063  1.0774 -1.5412
#>  0.0105  0.6940 -0.5086 -1.6552  0.1314  1.7843 -0.2984 -0.4991 -0.7196 -0.9581
#> -0.8219 -0.8282  1.2117  0.1088  1.2336  0.9698 -0.6830 -2.4649 -0.4814 -1.0124
#> -1.4868 -1.6977  2.8886  0.4825 -0.5261 -0.9333 -0.1885  0.9150  1.4861  0.3470
#>  2.0784  1.6903  1.6201  0.7490  1.1006  0.9867  0.9533 -1.5456  1.7646 -0.6083
#>  0.4671 -1.2967  0.2668  0.1321 -1.3650 -0.7321 -1.2496 -0.2394  1.2703 -0.1729
#>  0.9354 -0.4624 -1.3386 -0.8281  0.3321 -0.3384  0.2263  0.4059 -0.8847 -1.0366
#> -0.3279 -0.4362  0.2900  0.0336 -1.1202  1.3094  0.4105 -1.1267  0.5877  1.0013
#> -0.8926 -0.6634  0.0915  0.3228 -0.8834 -0.3990  1.4449  0.0013 -0.3356 -0.1204
#>  1.0580  0.6539 -1.3090 -1.2557 -1.1779  0.0316 -0.1235 -0.0330  0.6870  0.7225
#> -0.2790  0.4248 -1.2505  0.9233 -0.6800  2.6322 -0.3773 -0.1207  1.4497 -0.8025
#>  0.5085  1.1942 -1.5529 -1.2772  0.0629  0.9676 -0.4088  0.4205 -0.6426  1.1607
#>  0.4093 -0.7825  0.4087 -1.4188 -0.1706  2.1498 -0.0275 -0.1724  0.2193  0.4396
#> -0.5880 -1.7702  0.5365  0.5930  0.0302 -0.3829 -0.8705 -0.4673  0.7728  1.0554
#> -0.2665 -0.0152  0.0432  0.8097 -1.1610  1.1848  0.2444  0.4433  0.3537  0.2313
#> -2.1154 -0.3887 -1.2708  0.0780 -0.3710  0.9130 -0.1078 -0.7228  0.9022  0.3718
#> -0.3785  0.5087  1.4532  0.9228 -0.4643  1.9672 -0.9684 -0.7122 -1.2296 -0.2648
#> -0.9706 -0.7759 -0.8474  0.1972 -0.6867  0.7136 -0.8320  1.4975  0.0608 -0.0561
#> -1.1676  0.6206 -0.4694 -1.1602  1.0647  0.0072 -0.6808  0.3000 -0.8333  0.4908
#>  1.3865  2.4374  0.4777  1.3865 -0.6537 -1.1382 -2.2063  0.2438  1.7282 -0.5021
#>  0.2124 -1.1130 -1.5399  0.6478  1.7665 -0.0875  0.4412 -0.6171  1.7091 -0.8167
#> -0.7899  0.5837  0.0319  0.5363 -1.0599 -1.6567 -0.6267  1.9418 -1.0555 -0.1050
#> -0.3332  1.3884  1.2140 -0.1799 -0.4658  1.3026  0.1653 -1.4961  0.3917 -1.0111
#>  0.6459 -0.4833 -1.7017  0.2916 -0.9730  0.7529  0.4928  0.6558 -0.4878  0.9351
#>  1.1399  0.3072  1.8071 -1.4997  0.5512  1.7458  0.7608 -0.5900 -0.4898  0.3515
#>  2.1564 -0.7617 -0.5650 -0.2744 -1.2930  0.3237 -1.0288 -1.2339 -0.8803  0.1897
#> -0.3237  0.3279 -1.1501 -0.8051  0.4365  0.4222 -1.0230  0.1168  0.0543  0.1136
#> -1.4161  1.5154 -0.4477 -1.3895 -0.5763 -0.2409 -0.6134  1.5282  0.0856 -1.0532
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]