Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6-2.4499e-01 -6.1850e-01 -3.1045e-01 -7.8086e-01  5.0887e-01 -2.1168e+00
#> -3.0090e-01  8.6525e-01 -2.4270e-01 -6.8826e-01 -7.1994e-01 -4.3866e-01
#>  1.2015e+00 -6.4166e-02  4.7024e-01  1.7369e+00  1.6471e-01 -1.2959e+00
#> -2.4143e-01  1.2948e+00  5.5904e-01  9.2816e-02 -1.2527e+00 -8.7193e-01
#> -8.5704e-01 -1.1133e-01  1.9798e+00  8.4973e-01  8.8760e-02 -6.2844e-01
#>  1.9273e-02  2.5198e-01 -1.4070e+00 -4.1907e-01  1.6970e-01 -1.3639e+00
#>  9.6866e-02  1.0560e+00  1.3377e+00 -2.3640e+00 -9.6666e-01 -1.3778e+00
#> -7.0007e-01 -2.3928e-01 -1.0793e+00  2.8148e-01 -7.5753e-01 -2.0570e-01
#> -2.1330e+00 -6.3798e-01 -1.4689e+00 -5.0321e-01 -2.6716e-01 -1.5173e-01
#>  2.4471e+00 -1.9674e+00 -4.8465e-01 -1.1307e-01 -2.4027e-01 -9.7407e-01
#> -2.1709e+00 -3.1838e-01  8.4924e-02 -1.5897e+00  6.2333e-01  3.5005e-03
#> -5.6528e-01  1.6525e+00  3.1072e-01 -4.1029e-01 -8.5178e-01 -2.3326e-01
#> -6.5857e-01 -6.1402e-01  5.0508e-02  1.3608e+00  1.3095e+00 -5.7587e-01
#>  7.4608e-01  9.0161e-01  9.9207e-01 -8.0062e-01  2.3919e+00 -1.2621e-01
#>  4.2060e-01 -8.1315e-01 -1.3108e+00  3.9972e-01 -9.4670e-02 -3.4742e-01
#>  3.4412e-01 -7.0213e-01  6.5600e-01  1.6214e+00  1.9529e-01  2.9028e-01
#>  5.3490e-01  3.2820e+00  1.3697e+00  5.1273e-01 -7.0242e-01 -4.2164e-01
#> -1.4146e+00  1.6615e-01  1.1564e+00 -1.6573e-01 -5.7258e-01  3.7348e-01
#> -8.2603e-02  3.8227e-01 -4.7179e-01  1.1390e+00  2.2142e-01  6.1412e-01
#>  1.3732e+00 -7.6126e-01  1.0271e+00 -1.3165e+00  4.3654e-01 -8.4321e-02
#> -6.9917e-01 -8.6680e-01  1.2112e+00  4.3565e-01  1.1367e+00 -1.9612e+00
#> -6.7336e-01 -3.8482e-01  9.8698e-01  3.3436e-01  2.6046e-01  3.7059e-01
#> -1.0417e+00 -1.5250e+00  7.2976e-01  6.8369e-01  1.9405e-01  5.8371e-01
#> -7.1224e-01  1.1378e+00 -1.1214e-01 -1.6104e+00 -4.2249e-01  5.5575e-01
#> -5.8461e-01 -6.0525e-01 -3.6073e-01  6.9595e-01  5.4302e-01  2.9965e+00
#>  9.6505e-01 -9.5059e-01 -3.9870e-01 -1.8731e+00 -9.3762e-01 -2.3832e+00
#>  1.9686e+00 -4.7190e-01 -1.0963e+00 -6.1319e-01  1.5564e+00 -3.9320e-02
#>  9.8271e-01  8.8550e-02  9.0181e-01  7.4248e-01  5.6023e-01 -1.6184e+00
#> -2.0416e-01  3.3637e-01 -2.7730e-01  5.1184e-01  9.9674e-01 -6.4819e-01
#>  1.3811e+00  2.9102e-01 -5.5656e-01 -3.2016e-01 -4.4311e-01  1.6689e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]