Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7942 -0.9947 -0.5203 -1.6541 -1.0107  1.0497 -0.1256  0.6373 -0.4188 -0.9549
#> -0.0160  1.5910 -0.0508 -0.8195  0.3937  0.6724 -0.0011 -0.3161 -0.9918  0.0741
#>  0.3918  0.2130  1.0176  0.3713 -1.9094 -0.3273  0.3606  0.6798 -0.7655  0.8209
#> -0.5769 -0.1108  0.7239 -0.3277 -1.7429  0.6724 -2.5165 -0.5813 -1.1471 -1.0805
#>  0.1729  0.4261  0.7613  1.6436 -0.6291  1.3879  1.3293  0.9098  0.7707 -0.9204
#> -0.4596 -0.3526  1.3698 -0.4886 -0.3899 -1.0925  0.1462  1.1218  1.4454 -0.1239
#>  1.2565  0.5218 -1.8348 -0.4442  0.5072  1.9944 -0.3445 -1.8594  0.6577 -0.2665
#> -1.1085  1.8145 -0.8943  0.2898  0.5353  0.8042  0.0908  0.9607  0.5794 -0.8724
#>  0.2937  1.4475 -0.0230  1.1103  0.7136  0.1221 -1.8523 -0.2944 -0.3563  0.7878
#>  0.3815 -0.1673  0.1786  0.7168 -0.3166  1.0535 -1.1282  0.1593  1.0362  0.3687
#> -0.5062  0.5839  1.9742  0.4577  0.7735 -0.8009  0.4969 -0.7056  0.7859  0.2822
#>  0.2203  0.8917  0.3087 -0.1185  2.0822 -0.7627 -0.1336 -0.1333  0.6209 -0.2728
#>  0.4358 -0.6003  0.1581  1.8604 -0.5298 -1.0798 -0.3170  1.6213  0.3538 -0.0146
#>  1.2935  0.1297 -0.3121  0.7351  0.6230 -1.6034  0.8507  0.1822 -0.6777  0.1908
#> -0.8737 -0.8649  0.6505 -1.1013 -0.8135  2.2546 -0.1385  0.7886  0.7020  1.7171
#> -1.5485  0.5897  0.8472 -0.0863  1.6910  0.3644 -0.4625 -1.5641  0.0832  1.3932
#> -0.5263  0.5879  0.6820  0.0390 -2.0412  0.4313 -0.2223  1.2011 -0.9423 -0.2347
#>  0.1252 -0.4512 -0.4894  0.2894  2.0885  0.4158 -0.1556 -1.6916  0.4904 -0.6008
#> -2.0490 -0.3115  0.4772  0.5675 -0.4834 -0.0074  0.1746  1.0902  0.6913 -0.9578
#>  0.3901  1.0012  1.7042 -0.6967  0.9234  0.2449  0.1088  0.5291  0.2139  1.6833
#>  0.7014 -1.7182  0.3442 -0.1073 -0.6263 -0.6658 -0.5780 -1.0224 -1.2448  1.5007
#> -0.6922  0.2700 -0.4536 -0.3631 -1.7012  0.9281  0.8624  0.8924 -0.3390 -1.5765
#> -1.1712  1.4397  0.4250  1.2764 -0.6945  0.7510 -0.7553  0.0043 -0.7040 -0.8011
#> -2.0054 -1.2517 -0.3197  0.0797  1.0200 -0.4110 -0.4805  0.3961 -2.1996 -0.2986
#>  0.5458 -0.5456 -0.1600  0.2132 -1.2360 -0.6187  0.0460 -1.0370 -0.0859  0.7694
#> -0.5332 -2.2847  0.7264  0.1298  0.1839  0.0676  0.2973  0.7003  1.1633 -0.0276
#>  2.0036 -0.0764 -0.0933  0.1216 -0.3252 -0.2085  1.5957  1.0294 -0.6548 -1.0471
#>  0.2316 -1.1932  0.6402 -0.1636 -1.4937 -0.2899  0.0657 -2.2076  1.7044  1.1082
#>  0.7495 -1.0940  0.5936  0.9406  0.5847  0.6749  0.4219 -0.0350 -0.2944  1.4447
#> -1.9955 -0.4008 -0.1427  0.8122 -0.7220  1.3724  0.3845  0.4395  0.5842  0.9683
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]