Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10 0.5344 -1.1088 -1.4895  0.4322 -0.5890  0.1500  0.2214 -0.2012 -0.5406 -0.0695
#>  0.2862 -0.1990 -0.8028 -0.2234 -1.1598  0.1014  0.4385  0.5822  0.3823  0.7665
#> -0.5431  0.1197 -0.8786  0.5425  0.3165 -0.6045 -0.2354  0.5511  0.9152  0.4952
#> -0.3738 -0.3837 -0.7140 -0.5484  0.6044  0.6565  1.4990 -1.9579  0.5798  0.5512
#> -0.2317 -1.8156 -0.9288 -0.6464 -1.4210 -0.2584 -0.7502 -1.0416  0.0729  1.5367
#>  0.4679  0.0203 -0.9480  0.8460 -0.9607 -0.5955  0.1429 -0.4700 -0.5496 -0.1930
#>  0.4353 -1.0529 -1.4982 -0.6335 -0.9126  2.2660 -0.7529  0.6307 -0.9633  0.2494
#>  1.0178  0.9702  0.0803  1.1512 -1.0345 -0.0668 -1.1152 -1.8751 -0.5746 -1.0528
#>  0.9937  0.1321  1.4749  1.2485  1.6610 -1.9611  1.0366 -0.1646  1.1503 -0.8731
#> -0.4663 -0.3292  0.4192  3.0274  2.2827  0.3061 -0.0002  0.3636  1.0320  0.3970
#> -0.2302 -2.0125 -1.0585  1.4908 -0.2486  1.1199  0.6393  0.0762 -1.4258 -1.4975
#>  1.1918  0.4399  1.7374  1.1990  0.5327  0.7905  0.5101 -1.0095  2.2117 -0.9021
#>  1.7000 -1.1318  0.2071  0.4429 -0.4049  0.6003 -0.6245 -0.2174 -0.9311 -0.5057
#> -2.2745  0.9090  1.0931 -0.6431  0.8984  1.2812  0.9218 -0.0287 -1.4807  0.1597
#>  0.2405  0.3020 -0.6726 -0.8438  2.6003 -0.4983 -1.6037 -0.0899 -0.1289  1.1970
#> -0.3180 -0.4801 -1.7658 -1.1434  0.3249 -0.3922  0.8681 -0.0903  0.2642 -1.0958
#> -0.5909  0.4816  0.9101  1.3723  1.0619 -0.1049 -2.5438 -0.3927  0.3218 -0.2248
#>  0.6874 -0.8661 -0.8022  1.2145  1.4858  1.0177  2.3434 -1.0673  1.5294  1.2484
#> -1.5861  0.4552 -0.3601 -1.5904  0.7037 -0.8252 -0.8511  0.1451  0.9645  0.0770
#> -2.0474 -0.8586 -0.5945 -0.7912 -2.5196  1.7148 -0.6913 -2.1132  1.5265 -0.1888
#> -1.3509  0.8825 -1.3293 -0.0682  0.5129  0.2319  0.2570  0.2674  0.8834  0.3301
#>  0.8801 -0.7796  1.0042  0.1519 -0.7694  1.6389 -0.1426  0.5004 -0.4030 -0.1910
#>  0.3731 -0.5751 -0.2533 -0.4321 -0.6270  0.6773  1.4346  0.3329 -1.9132 -0.1929
#>  0.0093  0.4235 -0.3915 -0.2992 -0.0588  0.2015  0.1177  1.9111  0.0134 -0.1983
#> -0.0582 -1.9748 -1.2293 -0.8940  0.2536 -0.7570 -1.7984  0.1584 -0.3878 -2.6656
#> -0.6704  0.1694  0.9738  1.4057 -0.3312  0.1527 -2.2629  0.2606  1.1305  0.6123
#>  0.7866  0.2044 -0.6325  0.9213 -1.4295  1.3466  1.7918 -1.4834  0.0665  0.2864
#> -0.3021 -0.2000  1.1770 -1.2727  0.1572 -0.6952  2.3195  0.2689  0.5157 -0.1017
#> -0.7581 -0.8904  0.4489 -0.5078  1.1045 -1.2314  0.0304 -0.4325 -0.8163 -0.4027
#> -1.8508 -1.7187 -0.5961 -1.1793 -0.2980  0.5728 -1.6535  0.0005 -0.3973 -0.9023
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]