Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.4519 -0.2492 -1.5222 -0.3818  0.3669 -0.0696 -0.6610  0.5141  0.9594 -1.0774
#> -1.1968 -0.3797 -2.0371 -1.6899  0.4544  1.3268 -0.0966 -0.9680 -1.2009  0.0116
#> -0.4168 -0.9520  0.5526  1.2170 -0.3620 -0.0717  1.0407 -0.6610  1.9470 -1.4932
#> -0.0643  1.1769  1.3287 -1.6072 -0.8698 -1.3288 -1.7136 -0.4094 -0.0503 -0.4199
#> -0.1604  0.4387  1.4415  0.1434  0.2966 -0.9463  0.2604 -0.8538 -0.5589 -1.6661
#>  1.2954 -1.0045 -0.4514  1.1488  0.5351  1.6097 -0.8243  0.0104 -1.5251  1.7708
#>  2.0246  1.2806 -1.7124 -0.1302  0.5619  1.1203 -0.9855 -2.5801  1.3502  0.8767
#>  0.8046 -0.7329 -0.7399 -1.4219  1.6505  0.6053  0.5865  1.2288  0.5320  1.3920
#>  2.0766 -0.5535  2.8500 -0.0483  1.3022 -0.6258  1.4549  1.3801  0.3853  1.0885
#>  0.2851  0.2510 -0.6494  0.4766 -1.3557 -0.8670  0.6281  0.0716  1.5554 -1.4909
#> -2.4319 -0.8263 -1.5765 -2.3407  0.8031  1.4339  0.5510 -0.8467 -0.9435  1.1615
#>  1.9244  0.4675 -0.7645 -0.1755  0.2317 -2.2613  0.0364 -0.9287  1.1777 -0.6348
#> -0.8567 -0.6878  0.4411  1.4432  0.7422 -0.2764  0.3314 -0.8900 -0.4745  0.6655
#>  0.1372  0.3339  0.6698 -0.0100  0.5434 -1.0815 -0.5708 -0.6787 -0.3344 -0.7150
#> -0.2808 -0.4398  0.6172 -2.8697  1.3314  0.7503 -0.8458 -0.3646  0.7436 -0.6210
#>  0.0980  0.7934 -1.2561 -1.5609 -0.4147 -0.5877  0.5499 -1.0572  0.5354 -0.6144
#> -0.2443  0.5897  0.1669  0.9863 -0.3695 -0.3784  0.1725 -0.0014 -0.4786 -0.7148
#>  0.8926 -1.3710 -0.7762  0.8559  0.7110  0.9578 -1.7644 -0.0412 -1.3360  0.1092
#> -0.6945 -1.3978 -0.4436 -1.8114  1.6243 -0.9613 -0.1725  1.7280 -1.5733  1.4740
#> -0.9227 -0.4893 -0.5289  1.4192 -0.9123  0.1018 -0.1774 -0.6055  0.1518  1.4742
#>  0.7881 -0.4418 -0.4916  0.5716 -0.0040  0.3193  1.2818 -0.1453  0.9474 -1.1451
#>  1.5308  0.3635 -0.6754 -1.3612 -1.4424  0.8120 -2.1882 -0.9078  0.5554 -0.4049
#> -2.3153 -0.9270  0.5893  0.5298 -0.0194 -1.7999 -0.1886 -0.6985 -2.1621  1.5037
#>  0.1485 -1.0614 -0.2067 -1.0199 -0.4958 -0.3993  0.8296 -0.8150  0.1250 -1.2452
#>  0.6523  1.2895  0.3182 -0.2375  0.1782  0.6960 -0.3665  1.9043 -0.2393  1.2367
#> -0.8509  1.2266  0.5537 -0.7662  1.2200 -0.0407 -0.7706  1.5620  2.2673  0.6237
#>  0.0477 -0.0970 -0.4895 -0.9421 -0.9255 -1.4541  0.1554  0.1480  2.0307  1.4946
#>  0.6070 -0.9594  1.4170 -0.0017 -2.2886  0.7436 -0.1327 -1.1536 -1.6252  2.7740
#> -0.3052  0.9332  1.0697 -0.3537 -0.6553 -1.7362  0.0468  0.0141  0.5436 -1.2868
#>  1.0881 -1.2099 -0.9959  1.1764 -1.7488  0.2347 -0.9832 -0.4364  2.2397  1.0248
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]