Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.9295 -0.0179 -0.0312 -1.3216  0.0297  0.7353 -1.5919  0.1756  1.0239  1.1880
#> -0.8209 -0.2591  0.3118 -1.6476 -0.1150  1.2715 -0.2931 -0.7283 -1.2460 -0.9083
#> -0.0794  0.1338 -0.1800 -0.4036  1.4412 -0.1312  0.8168 -0.3334 -0.0431 -0.7393
#>  1.0889 -0.4591 -1.9138  1.4610 -0.3654  1.9395 -2.1749  0.5258 -1.5704 -1.2566
#>  0.4856 -0.6226  0.0255 -1.0815  1.1087 -0.6459  1.1723  0.7796  0.1150  0.5160
#>  1.1891 -0.4050 -0.4386  0.9419 -0.1534 -0.8078 -0.3268  0.4260  0.3550  0.2928
#> -0.5343  0.7674 -0.5726 -0.9150  0.0694  0.9978  1.0405 -0.6691 -0.5306 -1.0772
#>  0.6918 -0.8480 -1.6506  0.6527 -0.7139 -1.1559 -0.1832  0.2842 -1.0562 -0.1826
#>  0.2229  2.4111 -0.4520  1.1441 -0.1350  2.6859 -1.3701  1.3267  0.4968  0.6975
#>  0.4452 -0.4118 -0.2542 -1.6856  0.4938 -0.0068  0.1794 -1.2872 -0.5261 -0.3273
#> -1.3275 -1.2827 -1.4893  0.1817 -0.0553 -0.0060  1.0012 -0.1065 -2.3910  1.1536
#> -0.5824 -1.2571  1.4883  0.0369 -0.9527 -0.6116  0.0908 -1.1307  0.9962  0.2153
#> -0.7625 -0.3780  0.2483  0.9994  1.6090  0.6226  0.5599  0.6044 -0.1761 -0.7240
#>  0.2027 -0.0421 -0.2246  1.5939 -1.0368 -0.5169 -1.4535 -0.5804 -0.7194  1.2911
#>  0.7487 -0.8667  1.0207 -0.2880  0.1994 -0.9958  2.1850  1.0938  0.1561 -1.3421
#> -0.1505 -0.0828  0.4013  1.9843 -1.6880 -0.7123 -0.3862 -0.8567 -0.2759  0.1983
#>  0.8772 -1.1104 -2.2627  1.0909  0.4114  0.1419 -0.5698 -2.6872 -0.9436  1.5874
#> -0.0501 -0.8133 -1.5213  0.5304 -0.1255  2.3687  0.1381 -0.0270  0.4478 -1.4322
#>  0.6265 -1.8093 -1.0952  0.1933 -0.8026 -0.4398 -1.5638 -0.5247  1.7155  1.0198
#> -0.5979 -0.2763  1.0331 -0.7070 -0.3137 -0.0897 -0.8482  0.1995  0.4448 -0.5332
#>  1.7488  1.5301 -2.2662 -0.4972  0.7660 -0.3249  0.7180  0.8625  0.1815  1.6174
#> -1.6472  0.1818  1.3780 -0.2193 -0.3016 -0.6587  0.9822 -1.0818 -0.7933 -0.4239
#>  1.4008 -0.7657 -0.3981 -1.3337 -1.4770 -0.2478  0.9176  0.0832 -0.8935  0.6201
#>  1.3625  2.6819 -0.2443  0.8784 -1.3514 -0.4487  0.3676 -0.5087 -1.2044  0.8926
#>  1.1378 -1.6492 -0.7345  0.2879  1.9103 -1.2856 -1.3369  0.0058  0.5648  1.4030
#>  1.4030 -0.4380  1.4897  1.7020 -0.3994  0.1610 -1.1709 -0.3107 -0.4823 -0.0398
#>  1.3971  0.3372 -0.7548  1.0945 -0.2645 -0.9367  1.0864  0.7004  1.3315  0.5167
#>  0.3181  1.4644 -1.2131 -0.7951  1.3012 -0.6995  1.0468 -0.9656  0.4489  0.3981
#>  1.4899  2.1037  0.2009  1.5670 -0.2936 -0.7548 -0.2343 -2.6427  0.5917  0.1211
#>  1.1724 -2.0773  1.0481 -1.8759  0.0260  0.0046  0.4637 -1.1388  0.6866 -0.8275
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]