Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0926 -0.1265 -0.2943  0.1091  0.9860 -2.3439 -1.1358 -1.4944 -1.1202  2.0612
#> -0.0184  0.0098 -1.2849 -0.0579  0.3813  2.0602 -0.0711  0.6589  1.5199  0.0988
#> -0.6116 -2.4104  1.6584 -1.1423 -0.7482 -0.1899 -1.4640  0.6746  0.0255 -0.2357
#>  1.0369 -0.7942 -0.2648  1.5054  0.3901 -1.0208 -0.9046 -0.6972 -0.3617  0.1938
#>  0.3552 -1.3323  0.5653  0.0024 -0.1326 -1.3160 -1.4877 -0.6486 -0.7472 -0.1409
#>  0.5532 -0.0642  0.5279  0.6617  2.4259 -0.8335  0.3498 -0.6910  0.2820  1.3987
#>  0.1383 -1.5962  0.2976  0.0961  0.5948 -0.8128  0.6903 -0.5026  0.7903  0.8312
#> -0.8720 -0.1758 -0.4261 -0.7445 -1.4377  0.0145  1.1714 -0.0750 -0.6640  0.5741
#>  1.1155 -0.4102 -0.0835 -0.5966 -1.1612 -1.1283  3.0795 -1.2393 -0.1723 -1.0342
#>  0.9199 -0.7372  0.6462 -0.3136 -2.4339 -0.4436 -1.8455 -2.4385 -0.9175  0.6157
#> -0.4812  2.0678 -0.6722  0.2544  0.2009 -0.5456 -1.3006 -1.0030  0.0740  0.6438
#>  0.0167 -0.5906  0.5871  1.2426 -0.7188 -0.0201  0.4089  1.4431 -0.0334 -1.8267
#>  0.8190 -0.3751  0.0174 -0.5374  2.1589 -0.8217 -0.3263  1.7721  1.4610 -0.9450
#>  1.2844  0.4693 -0.6631  1.3968 -0.1870 -0.2283  0.6765 -0.0878 -1.7557  0.2824
#>  0.3175  1.8460  0.4085 -0.6418 -1.2598 -2.8991  0.6154 -0.1306 -0.6337 -0.8262
#>  0.3971  0.9660 -0.2683 -0.2994 -1.1039  0.9130 -0.4154  1.0875  0.5794  1.2578
#> -0.0898 -1.5525 -1.2525 -1.6316 -1.1410 -1.5361  0.6660 -0.1925 -0.5569  0.0751
#>  1.1118 -1.1760  1.6093  0.3238 -0.0842 -0.4319 -2.7608  0.5409  0.1715 -1.3053
#> -0.3325  0.1636 -0.1228  0.2597 -1.4157 -0.0510  2.2505  0.0708 -1.3195  1.0940
#>  0.0365  2.7575 -0.2102 -0.8445  0.5197 -1.8462 -0.5277  0.2242 -0.6565  0.0236
#>  0.1855  0.5585  1.3260 -0.3094 -1.0444 -0.2019 -0.8082 -2.7515 -1.7596  0.4141
#> -1.0929  0.7843 -1.4341 -0.9962  0.6890 -1.4130  1.0743 -0.2898  1.3525 -0.0698
#> -0.5661  3.0627  0.1626  0.1706 -0.0649 -1.2589 -0.0189 -0.1633  0.6274  0.1397
#>  0.0320 -0.1820 -0.9540  0.1890  2.4428 -1.0507  0.5936  1.0281  0.1711  0.1829
#>  0.3531  1.0300  0.0853 -0.1186 -1.0243  0.0113 -0.4897  0.4706  1.1211  0.9710
#> -0.8704 -0.1660 -0.0842 -0.1026 -0.3267 -1.7009 -1.3248 -1.0245  0.3126 -1.8296
#>  1.0189  1.0369  0.0576  0.0972  1.1983 -0.8388 -0.5151 -0.4405  0.1948  1.3508
#>  0.4787  0.1059  1.4529  0.7499 -0.1770 -0.4915 -0.0031 -1.1084 -2.4135  0.2718
#>  0.0533 -1.3965  0.6395 -2.0159 -0.3599  0.2386 -0.1854  0.1729 -0.5091  0.2834
#>  0.9157  1.1772  1.2395 -0.8222 -0.4962 -0.0685 -1.1046  1.2601 -0.6934  3.0228
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]