Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.1374 -0.5571 -0.1412 -0.5196  0.0169  0.2314 -0.6713 -0.5552 -2.0556 -0.1259
#>  0.9848  1.9562 -1.1230 -0.6892 -0.5701  1.4469  1.0994 -1.4411 -1.2438 -1.1589
#>  0.3958 -1.0570 -1.0152 -0.7527 -2.6658 -1.3460  1.1281 -0.9885  0.0193  0.1977
#> -0.6007 -0.5773 -1.1715 -0.0674  2.3989  0.9080  0.1621 -0.3217 -0.7800  1.1505
#>  1.0572  0.3385 -1.1920 -0.5268  0.0960 -0.5885  0.6648 -0.2578  3.2506 -0.3715
#> -0.3747  2.7915 -0.4916 -1.1381 -0.8093 -0.7404  0.4450 -0.5024  0.4312  1.3534
#> -0.0630  0.0401  0.9543  1.1424  0.0659  1.9409 -1.3921  0.2404 -0.0455 -0.0545
#>  1.2900 -0.1186  1.7323  1.8775  0.1570 -0.2184  1.8923 -0.1872 -0.6696  0.0045
#> -0.7741 -0.1009  0.9772  1.1460  1.0501  0.0036  0.1687  2.0438  0.6636 -0.6422
#> -0.3953 -0.5450 -1.0327  1.4087  0.9348 -0.7028 -0.1355  1.5014  0.1605  0.8036
#>  1.3848  0.6569 -0.2151  0.6344 -0.2038  0.5121  2.5281 -0.3681  0.5929 -0.4501
#>  1.1692  0.9022 -0.7794  1.1292 -0.2001  1.0223  0.2629 -1.3782 -0.4632  0.6884
#> -2.2839  0.5691 -0.0823  0.0212 -1.3250 -0.7598 -0.2414 -0.2417 -2.5045  1.3903
#>  0.5100 -0.5721  0.0289 -0.2771 -0.3877 -0.1598  0.4346  0.3559 -1.2423 -0.2206
#>  0.6879 -0.6030  0.1273  1.0240 -1.3083 -0.2880 -0.3939 -1.1750 -0.7541  0.1269
#> -0.7082  0.3259  0.5593 -0.4335 -0.0426  0.2958 -1.3951  1.7595  0.0849  1.1203
#> -0.5369  0.5384 -0.7813 -0.5596  0.1306  1.3385  0.6446  0.5360  0.2894 -0.4015
#>  1.0475  1.2849 -0.0075 -0.1153  0.4173 -0.7273 -0.4401 -0.1264 -0.4359  0.1482
#> -0.6991  0.9798  0.7276  0.9927  0.5560  0.5710  0.5909  1.1207 -0.3956  0.5047
#>  0.9481  1.6588  0.6990  0.5055  1.8303 -0.2376  0.4015  0.0550  0.0480  1.1600
#> -0.8649 -0.0515 -0.1331 -0.0715 -0.8927 -0.6199 -0.3549  0.7337 -0.3992 -0.1608
#>  2.2552 -0.7697  0.7387  0.3767  1.9244 -2.4361 -0.6582 -0.1442  0.6944  0.0976
#>  0.5007  0.0133  0.1294 -0.5464 -0.1189  3.0275  1.1408 -1.2705 -0.5428  1.0195
#> -1.1134 -0.1933  1.1874 -0.9552  0.4652 -1.2636  0.0112 -0.2836 -1.7545  0.9419
#>  0.3913 -1.4168  0.3389  0.8272 -1.6144  0.0650  0.6654 -0.5081  0.8752  0.1014
#>  0.7576  1.0887 -0.0941 -0.7069  0.5549  0.2409 -0.1204  0.1792 -0.5295 -1.0002
#> -0.3922  0.3468 -0.5501  0.7445  0.3734  1.3006  0.1428 -2.8207 -1.3762  1.3739
#> -0.4678 -0.3852  2.4652 -1.1410 -0.2422 -0.3143 -0.1462 -0.2989 -1.6670  1.4955
#> -1.4202 -0.3079  2.1588  0.1508 -1.4243 -0.5467  0.0626  0.3107  0.8976  0.8616
#>  0.0931 -1.7544  0.5625 -0.6544  0.0307  2.0437 -0.5633  0.7489  0.2335  1.3476
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]