Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 2.1035 -0.6875  0.1368 -0.5435  0.5650  1.1774 -0.7015  0.6582 -0.3770 -1.0648
#>  0.1869  1.3999  1.4234  2.9909 -1.4209 -0.4524 -2.3878  0.3547  1.6890 -0.1898
#>  0.0833 -0.0714 -0.0405  0.9168 -0.8398 -0.2157  0.1080 -0.1798  1.0672  0.4988
#> -0.4006 -1.6010  0.4986  0.2300  0.3255  0.2538  0.3654 -0.9721  0.3454  0.0486
#>  1.2477 -1.0453 -0.1954 -0.0052  1.1161  1.1766 -1.4515  0.4320  0.1827 -0.6549
#> -1.6933 -0.3622 -0.4711 -0.7355  0.1895  0.9715 -1.0716  0.1896  0.7206 -0.7645
#>  0.1578 -1.6287 -0.2025  0.7297 -0.4021 -0.5255 -1.2232  1.1503 -0.0818 -1.0987
#>  2.0274  1.7669 -0.7546 -1.8855 -0.1982  0.8875  0.2125  0.1502 -0.5599 -0.9046
#> -0.3979  0.8960  0.8974 -0.7098 -0.4244  0.2547  0.3436 -0.3145 -1.7873  0.1191
#> -1.1889  0.0281 -0.5594 -1.2788  0.8920  2.0011  0.1775  0.2078  1.0446 -0.4889
#> -0.8445 -0.7774  1.4361  0.1147 -0.6163  1.3982 -1.9578  0.9084 -0.6908  1.0576
#>  0.2775  0.0532 -0.3173 -1.2605  0.1003  0.0602 -0.1847 -1.3665 -2.3578  0.0741
#>  0.0093 -0.9263  0.5321 -0.5655  0.7631 -0.1866 -0.2116 -1.5267  1.3141 -0.4625
#> -1.5836 -2.1203  0.1325 -2.9032 -0.5195  0.4779  1.9599  0.4625 -1.3467 -0.0911
#> -0.7144  1.2037  1.8990  0.2837 -1.8460 -1.1960 -0.7144  0.1899  0.4289 -0.5978
#>  0.2092  0.5317 -1.3772  0.7842  1.5504  0.5747  1.4374  1.0031 -0.2785  0.0690
#> -0.2095 -1.9928 -1.0181 -0.4073  1.3272 -0.2611 -0.9915 -0.2577  1.3135 -0.1128
#> -0.7913  0.8379 -0.2701 -1.6994  0.2634  0.9507 -1.4090 -0.9181 -0.2271 -1.9081
#>  0.7596 -0.6633 -2.2616 -2.4802 -0.0986 -0.3176  0.3220 -0.0509 -0.6139  0.5951
#>  0.1174  0.0916 -0.7257  1.0990 -0.1894  1.9372 -0.4375 -0.1360  0.1925  0.7667
#> -0.0256  0.2902  1.3636  1.5857 -1.4290  2.1502  0.4968 -1.1694 -1.9688 -1.8193
#>  0.6378  1.4944 -2.4937 -0.4713 -0.3252 -0.6071 -0.5846  0.1693 -0.6436  2.1224
#>  1.9246 -0.0155 -0.6110  0.2418 -1.0962  0.0774  1.2826  0.8919  1.0546  0.3154
#> -0.7195 -0.2354 -0.5550  0.3373 -0.3688  0.8477  0.4304 -0.3904 -2.0686  1.6144
#>  0.0508  1.7121  0.1457 -0.7508  1.4664 -0.1345 -1.5872  1.3601  0.5173  0.3049
#>  0.2157 -0.2245 -0.4175  0.3921 -0.4622  0.4258 -0.7128 -0.0528  0.4056  0.0333
#> -0.5850  2.3850  0.1503  1.8421  0.7945 -0.9306  2.4465 -0.9917 -1.0207  0.0247
#>  0.1204  1.3011 -1.9604 -1.3125  1.1355  1.0303  2.9500  1.0036  1.1144  0.5295
#>  1.5826  0.1502 -0.9726  0.0693 -0.5641  0.6078 -0.6940 -1.0279 -1.9630  0.8159
#> -0.6910 -0.7851 -1.4751  0.0167 -0.1048  1.6909  1.2129  1.0383  0.9325  0.0694
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]