Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-1.6496 -0.3797  2.9840  0.1970 -0.8891 -1.0641 -2.0308 -0.9809 -0.0843  0.8322
#> -0.2781 -0.5080  1.4679  0.6464  1.0112  0.3893  0.0466 -1.4777 -0.7124  0.2506
#> -1.3030 -0.3966  1.2593 -0.2254 -0.1154 -0.9655 -0.3415 -1.3460 -1.3942  0.0762
#> -0.3689  0.1460  0.5818  1.3181  0.9135 -0.8680 -0.2975 -0.2452 -0.2707 -0.4975
#>  0.0908  2.5931 -0.7008  0.5269 -1.5886 -0.1165 -1.6593 -0.4999 -1.6430 -1.7785
#>  0.7616  0.1490 -0.9422  0.6242 -0.7558 -1.3798 -0.2142 -0.7830 -1.0514  1.6396
#> -1.3715  0.8454  1.2872 -0.8259  2.2298 -0.8923  0.3353  1.8668 -0.5299 -2.1285
#> -1.3152  0.7373 -0.5431 -0.0451  0.6162 -1.7014  0.6840  0.4972 -0.2788  0.8116
#> -0.5079  1.1993  0.0262 -0.7685 -1.3407 -0.6644  0.7941 -0.2520 -0.2402  0.4925
#> -1.0851  1.0790  0.1024 -1.1148 -0.0571  0.4253 -1.2853 -0.3600 -0.1123 -0.9172
#> -0.5656  0.2100 -0.4228  0.1798  1.0074  0.9236  0.2511  0.1923 -1.6975 -1.6746
#> -1.0261  1.3827  0.8458  1.3969  0.3959 -0.9189  1.6691  1.1654 -2.1333  0.5372
#> -0.5064  0.0611 -0.7588  1.6392 -0.2732  1.3824  0.0312 -1.4753  0.0377 -0.2683
#>  0.5058  1.4489  0.7836 -0.6168  0.4447 -0.3883  0.0069 -0.2938 -0.4785  2.2040
#> -1.0350  1.3378  1.8169  0.4128  0.2099 -1.9599  1.2077 -0.5905  0.5112  1.5259
#>  0.3326 -0.6441 -0.9078 -1.5382 -1.4313  0.2954 -0.3694  0.4649  1.8460  0.4335
#>  1.6848  1.7090  0.5588  0.5092  0.2682  0.2721  0.5638  0.4548 -0.3931  0.3872
#> -0.6320 -0.4144 -0.6788 -1.5025 -1.0382 -0.6581  0.1330 -0.1195  0.4222 -1.3601
#>  0.2858 -0.3983  1.6261  0.8762 -0.4014 -0.4148 -0.2042 -0.8654 -0.2567 -0.0030
#>  1.4121  0.0532  1.3784 -0.4507 -1.5576 -0.3263 -1.3759  0.6475 -0.3563  0.0140
#>  0.1755  0.1370  1.4288 -1.4452  1.4186 -0.1790  0.6597 -2.9900 -0.4741  0.4269
#>  0.5192 -0.4234  0.2412  0.1139  1.4910  2.0458  0.3599 -2.0420 -1.1842  0.7251
#>  0.4511 -1.0641 -0.6864  0.1654  0.4889  0.2692 -1.0458  0.4716  1.2305 -1.4798
#>  2.3100  0.8794 -0.2634 -1.3025  0.8559 -0.0920  1.0006  0.0851  0.7882 -1.2663
#>  0.9363 -0.3275 -0.6348 -0.2481 -1.2766 -0.6891 -0.3190  1.5084 -1.6624 -0.0027
#>  1.0439 -2.9685 -0.8646  0.8605 -0.5738 -1.1896 -1.5787 -0.2607 -0.0596 -0.7581
#>  0.8393  2.0908 -1.6224  0.1447  1.0089  0.2960 -0.6076  1.6315 -3.2831  0.5238
#> -1.2015  0.6806  1.9808 -0.1718 -0.4586  0.5085 -1.8552 -0.3336  0.7610 -2.0429
#>  1.7512  0.3211 -0.9241  0.5116  0.1532 -0.5808 -0.7588  0.6463 -2.1692 -1.7921
#> -0.6081  0.7109 -0.6131  0.5271 -1.3172 -0.0535 -1.2568  0.4513  1.7319 -0.6052
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]