Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.2118 -0.2679 -0.7442  0.9855 -0.0053 -0.6424  0.9718  0.2343 -1.0887 -0.8661
#> -0.0784 -1.0299 -0.4441 -1.4726  1.7128  0.7064 -1.9221  0.2160 -0.2246  0.4045
#> -0.3064  0.5596 -0.2321  0.7167  0.0971 -0.1173 -0.4395 -1.0767 -0.3539 -0.1773
#> -0.7518 -0.7531 -0.4320 -1.0029  0.3605 -0.2808 -0.3358 -2.2168  1.4748  0.8306
#> -0.6225  0.3899 -0.4569  0.2050 -0.6502 -0.6046 -0.5903 -0.7371 -1.3976 -0.1288
#>  1.0828 -0.5374  0.4671 -0.9656 -0.0382 -0.3372 -2.6428 -0.1159  0.3983  0.5906
#>  0.7797  0.5000  0.6272 -0.4107  0.9834 -0.6403 -2.2060 -1.0073 -1.0076 -0.0148
#> -0.2206 -0.6920 -0.3098  0.3447  0.0473  0.2634 -0.0623  0.8585  0.2827  0.1996
#> -0.6633 -0.3428  0.8803  1.5273  1.2156 -0.8773 -1.0090 -0.7641 -1.6932 -1.5590
#> -2.1759 -1.3236 -0.8367  0.6041  2.4521  1.3534 -0.7188  0.9532  0.4860 -0.2868
#> -1.0644 -0.1604  0.0351  0.8126 -0.8221  0.7992 -1.6516  0.0303  0.0373  0.3799
#>  0.2289  0.2828  1.2168  0.4519  1.2735 -1.5945  0.9795  2.8086 -0.1262  0.9128
#>  0.2337 -0.1622  0.6486  0.5699  0.6096 -0.8190 -1.3557  0.0534 -2.4300  0.6743
#>  0.1641 -1.0334 -2.4699  0.1731  0.0815 -0.0035 -0.0913 -1.6085  1.7116 -1.0876
#> -1.2608  1.6535  0.0520  0.2811 -0.9835  1.7460 -0.8400 -0.3111  0.2483  0.5492
#> -0.9791 -0.6456  0.0199 -2.7392 -0.6484 -0.0247 -0.6249 -1.0002 -0.7260 -1.3720
#> -0.6133 -1.6770  0.7831  0.6109  0.4997  0.1261  1.4577 -0.9350  0.7781  0.4218
#> -0.8029 -2.0224  0.4337 -0.8861  1.6989  0.1777  0.1768  0.4543  0.2102 -0.0475
#> -0.0289 -0.0202 -0.1323 -0.6701  1.0771  0.2481  1.1048 -0.0973 -0.1589 -0.1374
#>  0.4395 -1.0745 -0.0424  0.0879 -0.9353  0.7708 -0.6028  0.2524 -0.6367  0.8322
#>  1.6209  0.0816  1.3389  0.2129  1.0572 -0.0890 -0.5241 -0.2927 -0.2140 -2.0481
#>  0.9578  0.4753 -0.3875  0.4833 -1.1542  1.2980  0.6745  0.1810  0.0035  0.5351
#> -0.8608  0.8722  1.4809 -0.1023 -0.4051  0.4451 -0.5443  0.4954  0.0697 -0.7672
#> -0.9046  0.1151  0.7572  0.5390  0.0558 -1.3673  0.0680  1.1859  0.7934  0.2190
#> -1.0585  0.9721  1.0557 -0.0263  0.2801 -0.3891 -0.7873 -1.7212  0.3423  0.8157
#>  1.9572  0.9984 -0.5058  0.6490  2.3172 -0.1606 -1.8595 -2.5954 -0.0514 -1.1352
#>  1.2108  2.3255  1.0027  0.8004 -0.4045 -1.8933 -1.1727  0.8020  0.1623  0.7675
#> -0.8871  0.7153  0.9501 -0.3093  0.0387 -0.4659  0.6012  0.6673  0.4676 -0.6345
#> -0.2258 -1.1740  0.2544 -2.1952  0.9190  0.1747 -0.9177  0.0246  0.2006  0.1049
#>  0.4868 -1.5699  1.1449 -0.4785  0.2150  0.7893  2.0217 -1.2318 -0.8980  0.4061
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]