Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.3852  0.4841 -0.7555  1.4200 -0.3972  0.0575 -0.0742  0.9424  1.6513  0.0706
#>  0.8745  1.1329  0.3473  0.6537  1.3168 -1.6776  1.3826  0.3045 -0.1378  1.9417
#>  1.2907 -1.6654 -0.9814  0.7428  1.0435 -0.3421 -1.5210  0.7295 -0.7341 -1.5800
#> -1.1106  0.6460 -0.8337  0.1126 -0.0102  0.5825 -2.0011  0.8893 -0.6016 -0.0820
#>  1.2855  0.9039  0.2499 -1.1571 -0.3990 -0.8276  0.0417  0.9281  1.1819  0.2584
#> -0.5749 -1.1990 -0.4439 -1.2099 -0.2522  2.7553 -0.0781 -0.8917 -0.8796 -0.5097
#>  1.6950 -1.3080 -1.2577  0.0284  1.3476  1.3870  0.1578 -0.3128 -0.9123  0.9381
#>  0.2393  0.1032 -1.1776 -1.1604  0.2575 -1.4911  0.7896  0.2346  0.0310 -0.2963
#>  0.1514  1.0745  0.5369 -0.5480 -0.4825  0.3089  1.4064 -0.5597 -0.7748 -0.6822
#>  0.0728  1.0781  0.2083 -0.6178 -0.8795 -0.6236  1.7430 -0.1627 -0.0296  0.1553
#>  0.1942  0.3851 -0.9047  1.9163 -1.1628  0.2431 -0.3826  2.2121 -0.9666 -1.1010
#>  0.5159 -0.8470 -0.6692 -0.8603 -2.5556 -0.0431 -0.5883  0.4877 -0.6925  0.4367
#> -0.4114  0.1138 -2.0661 -0.7188 -0.3388  0.5125  0.6858 -0.8643 -1.6132 -2.4893
#>  1.6463  1.3544 -0.5204  1.1035  0.3497  1.3629  0.3379  0.5628  0.5784 -0.5750
#> -1.6745 -0.7432 -0.9758  1.7289 -0.3519 -0.5970  0.4057 -0.0653 -0.7511  0.7788
#>  1.4376  1.0385  1.1009  1.3467 -0.1741  2.2683 -0.3162 -0.7671  0.5145 -0.3250
#> -0.0965  0.3363 -1.1096 -2.1234  1.3136  0.3696 -2.7477  0.1731 -0.3471 -0.7877
#> -0.3346  0.0643 -1.3770 -0.2035 -0.3216 -0.8373  0.0498 -1.1270  0.2267  0.8374
#> -0.9745 -0.1618 -0.0766  0.4809  0.2779 -0.1381 -0.8858 -0.3270 -1.0099 -1.3611
#>  0.7828 -0.4289  1.5692 -0.5718 -1.9741 -0.4274 -1.6567  0.4485 -0.2136 -0.5259
#>  1.3649  0.9816  0.7193 -0.1390  0.0587 -0.5472 -0.1214  1.8237  2.0462  1.2168
#>  0.8043  1.4372  0.9466 -1.1499  2.1714  1.0567 -0.4027  0.5595  1.7443 -0.1701
#> -1.3919 -0.2194 -0.4793  1.3702 -0.4147  1.6178  3.0318 -0.1285 -1.1060 -1.8757
#>  2.3416  0.6238  0.8674 -2.0313  0.3750  1.4209  0.7767  0.0193  0.7120 -0.5672
#>  0.2267 -1.1135  1.6180 -0.4976 -0.2532 -0.1886 -0.8261  1.8324  1.4436 -0.1101
#> -1.7158  0.9751 -0.8445  0.7287 -0.8805  0.8302 -2.0371 -0.4051  1.6265 -1.1680
#>  0.0610  1.0970 -2.6850 -0.4521  1.7627  1.3688  1.0727 -1.1411 -0.3522 -0.8872
#> -0.2799 -0.7119 -0.5107  0.0501 -0.2092 -2.3259 -0.5052 -0.7198 -1.1199  1.3712
#>  0.7925  0.1307 -0.2734  0.4558  1.3458 -1.0156  0.1028  0.5766 -0.2934 -2.0947
#> -0.1510  0.3539  1.9871 -0.3737 -0.8053  0.9483  0.4438  1.1439 -0.1463  1.5309
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]