Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.1514  0.8848  1.1180 -0.0120  1.6081  0.5349  0.3676 -0.0489  0.5765 -1.5594
#>  2.2272  0.5383 -0.4330 -0.3156 -0.6036 -0.6932  0.1263 -0.5250  0.5089  0.4622
#> -0.1207  0.6575  0.4651  1.3747 -0.2380  0.3139  0.0791 -0.7604 -1.6369 -0.8014
#> -1.3241  0.9133  0.1552  0.7268 -1.0328 -0.2899  0.0690 -0.6818  0.6755  0.5567
#> -2.1242  1.2928 -0.5028  0.3897  0.4648  0.3277  0.0504  0.9385  1.1689  0.0131
#> -2.1480  0.0408  0.6034 -0.1208 -0.7150  2.6116 -0.4392  0.1405  0.1326  1.2824
#>  0.3990 -0.5271  1.0670  0.3932 -0.3014 -1.5916  1.2516 -0.2437 -0.9532 -1.1985
#> -1.2483 -1.2648  0.8780  0.2184 -0.1045  0.8364  0.6751  0.0684 -0.2354  0.1775
#>  0.1373 -0.5244 -1.1852  2.0378 -0.8222 -1.0114 -0.6272 -1.3992 -0.3097 -0.0306
#> -0.3499  0.1780  0.5944 -0.9481 -0.5638  1.7291 -1.3858  0.8563  0.1566  0.7144
#>  1.2208 -0.1541  0.9750 -1.0477  1.5259  0.9870 -1.5852  2.5315  1.3781  0.3904
#> -1.7995 -0.8089 -2.0391  0.4703 -0.8840 -0.1721 -0.1055  0.7756 -2.2150  0.7778
#> -1.6091 -0.6402 -0.1388  0.3196 -1.0384  1.3721 -2.0743  1.0719  0.4560  1.4642
#> -0.7022  1.7846 -0.3196  0.6546 -0.5915 -0.4968  1.0269  1.1945 -0.5160  1.8306
#> -0.4941  1.0292 -0.0752  0.5309  1.2259 -1.3406  0.1005  0.4151  1.0046  0.4981
#>  0.4663 -0.1920  0.0677  0.9866  0.5960  0.2545 -0.4996 -0.7455  1.5677  1.5108
#> -0.5632  0.1902  0.2003  0.7005  0.2141  0.6171  0.3138 -2.5720 -1.2050 -1.0678
#> -0.0555  1.4967  1.3836  0.4886 -0.4838  0.0406 -1.6212  0.0462 -0.6314 -0.4192
#>  0.6332  0.7927  0.0243 -0.0559  0.1463  1.4031 -1.0603 -0.2129  0.8846 -0.9967
#>  0.1012 -1.0356 -0.9750  0.8870 -1.6413  1.3080 -0.6307 -0.2467 -0.4069 -2.0404
#>  1.7233  2.0542  1.5189  0.3419  1.5490 -1.3009 -0.0887  0.8115 -0.3483 -0.9156
#> -1.7073  0.4093 -0.2665  0.8254  0.1979  2.8391  1.3467  0.1045  1.4666  1.2326
#>  0.1048 -2.0056  0.0354 -0.7092  0.1382  0.6541  0.3829 -0.5531  0.3630 -1.3920
#>  0.8275  1.4665 -0.7575  0.3354  0.7886  0.2724 -1.7258  0.7446  0.2099  0.3211
#> -0.9926 -0.1178 -0.7285 -0.8915  1.0889  0.4665 -1.0261  1.1930 -0.2690  1.0371
#>  0.9605  1.1391 -0.2835  0.7064  0.2294  0.1903  0.9881  0.9642  1.0567  0.1930
#> -1.8385 -0.4871  0.9696 -0.9825  0.3145 -1.2506 -1.0375 -1.0387  0.7485  0.9986
#> -0.5113 -0.4661  0.0342  0.3768 -0.5009  1.5788 -0.4674 -0.4091  0.4325  0.9358
#> -0.8467 -0.2597 -1.7281  0.7392 -0.7951 -2.8434 -2.1096  0.2287 -0.5829 -0.7286
#>  0.3332  0.6415 -1.3793  1.5914  0.1285 -0.0791  0.6550 -1.3512  0.3532 -1.4371
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]