Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.7521  0.0197 -1.1530 -0.0022  2.4894  0.6661  0.8070 -1.0902 -0.3335  0.4632
#> -0.7728  0.9188  0.9679 -1.1829  0.7331  1.3624 -1.7126 -0.3064 -1.2298 -0.0925
#>  2.2170  1.3296  0.5836 -1.8377  1.6864 -0.2238  0.4580 -0.2040 -0.0210 -1.6282
#> -0.1458  1.5735 -0.4717 -1.0447 -1.5662  0.0580  0.5022 -0.4143  0.8369  0.0893
#>  0.1515  0.7838 -0.2656  2.1093  0.6907  0.7329 -0.2348  0.2183  0.8510  2.2716
#>  0.8343 -0.3532 -1.0537 -1.4575  2.0240  1.0732 -1.8328  0.5404 -0.2610  0.3886
#> -0.0367 -0.6923  0.6762  0.2128  0.2261  0.6402 -0.3860  0.1657 -1.1684 -0.7807
#>  1.6574 -0.6174  0.2984  0.2657  0.0694 -0.3895  1.5818  1.6159 -0.2919  1.8881
#>  1.3988  1.7521 -0.6309  1.9882  0.9560  0.3958 -0.5523 -0.1298 -1.4582  1.0486
#> -1.0979  0.1877 -1.7367  1.4894 -1.7197  0.5381 -1.7280 -1.5129  0.6063  0.1009
#> -0.5733  2.3498  1.1968  0.4349 -0.7470  0.2933  0.1503  0.6493 -0.7746 -0.9729
#> -0.2081  0.4028  1.3235  0.2806  0.5208  1.6533  0.2985  0.3103  0.1226 -0.5210
#>  0.6102 -1.3394  0.6192  1.2214 -1.1599 -0.4779  1.0311  1.3009 -0.5055 -1.3591
#>  0.4079 -0.2872  0.2359  0.2889 -0.7317 -0.5904 -0.0609 -0.5303  0.6615  0.3637
#> -0.9542 -0.5746  0.3095 -1.4989  0.4951  1.6965  0.0940 -0.5747 -0.1266 -0.6794
#>  0.3308  0.1366  0.1953 -0.7162 -0.6142 -0.4334  1.3111  0.1516  1.5567 -0.5545
#>  0.1923 -0.1019 -1.9240 -0.5760 -1.0056  0.0492 -2.3477 -0.3650  0.6921 -1.0369
#> -0.3581 -0.8744  1.0620  1.2089 -0.7834  0.6764 -1.3207  1.7511  0.2829  2.1315
#> -0.3555  0.4487 -0.4607 -0.5182 -0.9462  0.3701 -0.2360  0.1330 -0.3111  1.6463
#>  0.2100  0.1001  1.0713 -0.6678 -0.3932 -0.5775 -1.1239 -1.1821  0.8521 -2.1926
#> -0.9125 -0.9818  1.6000  2.5355  0.3155 -0.5526 -0.2882 -0.5448  0.3938  0.1947
#> -0.7986  1.0712 -0.9075 -0.0644 -0.3923 -0.9272 -0.7972 -0.2103 -1.4268  0.4670
#> -0.2267  1.8596 -0.3918  0.9257 -0.4529  1.0075  1.2902  0.8391 -1.0261 -0.3745
#>  0.5267 -1.9515 -0.8773  0.9049 -0.0343 -0.7361 -0.1275  1.0837  1.0139 -0.7849
#>  1.1776 -0.4580 -0.2975  0.9167  0.4323 -1.8167 -0.3852 -0.7982 -0.4495  1.3804
#>  0.2554  1.3631 -1.6238 -0.7140 -0.2244  0.4871  0.4076 -1.0623  0.6834 -0.9703
#> -0.2658 -0.0219 -0.2400  0.7974  0.6708  0.2206  0.7804  1.5185  0.1029  1.0097
#> -0.3844  0.4805 -0.9497 -0.5650 -1.5412  1.1204 -0.2007 -0.0133 -0.7188  1.7329
#>  0.8333  0.6350 -1.7567  0.5082  2.6874 -1.7402  0.8580  0.8060 -0.4778  0.5319
#> -1.8218  0.1762  0.1252  0.2925  0.6714 -0.7349 -0.2028 -1.3205  1.2936 -0.0483
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]