Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7377 -0.8904 -0.8347 -0.1021 -0.8986  2.2675  1.5511  2.7033  0.4823 -0.1955
#> -0.6170  1.5445  0.2505 -1.7790  1.6127 -1.1422 -0.7806 -0.6836 -0.7567 -0.6967
#>  0.7948 -0.8518 -0.8757  1.1218  1.1499  0.4564 -0.9146 -1.1676 -1.0826  0.6988
#>  0.8002 -0.1344  0.8895  1.4977  1.1135  0.8862 -0.0240  0.2819  0.3954 -1.5599
#> -0.6900  0.6973 -0.2641 -0.4940 -0.9118 -0.4238 -1.8751 -0.6007 -1.0095  0.0655
#> -0.3854  0.0567 -0.1826 -0.5250 -0.2849  0.5427  0.8575  0.3930  0.4426 -1.7587
#>  1.0100 -0.8263 -1.2185  1.2050  0.9933 -0.5485 -0.6997  1.6863 -0.9482  0.1115
#>  0.4900  1.1963 -0.2499 -0.2860 -0.5915 -1.3880 -0.4277 -1.3046 -0.1418 -0.4846
#> -0.2451 -0.2956 -1.5844  1.2174 -0.1938 -0.8036 -0.3816  2.1237  0.4182 -0.4362
#> -0.0741  0.9943 -0.9971  1.1402  1.6430  0.7336  1.3049 -0.7672  0.0960  0.7913
#>  0.8829  0.6094  1.5881  0.4776  1.4416  1.9452  0.3075 -1.0655 -0.2605  0.2315
#>  0.7038  0.3172 -1.6056 -0.8386 -0.3507  0.3566 -1.2013  1.7591 -1.1529 -1.3438
#>  0.4447 -1.4767 -0.5087  0.8355 -0.4009  1.1481 -0.8007 -0.4991  1.3290 -0.2010
#>  0.0561 -2.4649 -0.0630 -0.2866 -1.4080  1.2670 -1.1009 -0.6396 -1.0658 -0.6126
#>  0.2262  0.3752  0.4880 -2.3046  1.2266 -0.1013 -0.8769  0.7084 -0.9262 -0.1653
#> -0.3223  0.4483 -0.0722 -0.1374 -1.4874 -0.2925 -1.2084  1.4349 -1.9662  0.3459
#> -0.0553  0.5806  2.4403  1.0068 -0.4851  0.7111  1.4737  0.3168 -0.0454  0.5554
#>  1.1696  1.3988  0.3056 -0.1046  1.5739  2.8028  1.3916  0.5586 -0.3307 -2.2033
#> -1.3365 -0.9196 -0.0848 -0.4079 -1.9732 -0.6517  0.5481 -0.7080  1.1436 -0.0333
#>  0.0263 -0.3432  0.4114  0.4201  0.5920  1.3045  0.4916  0.0052  1.4131 -0.2317
#>  0.8052  0.4818  0.6287 -0.4775 -0.1658 -1.2020 -1.2310  0.9003 -0.8268  0.2043
#> -1.4410 -0.3175 -1.5584  0.3554  2.6513 -0.0142  0.0315 -0.7284 -0.8532 -0.1603
#> -0.9215 -1.1585 -0.5935  0.0984 -2.4866 -1.2281 -0.5971 -0.6165 -0.4082  0.0417
#>  0.2625  1.7040 -0.1957 -1.7691 -0.3693  0.0856 -0.1348  0.0619  0.2369 -1.9775
#>  0.2445  1.2498 -1.2289 -1.1658  1.4763 -0.0612 -0.4990 -1.0064 -1.5602 -0.2263
#> -1.7413 -0.1812 -0.2736 -0.4910 -0.3941 -0.0754  1.4912  0.6659  0.1856  0.7144
#> -0.9461 -0.3922  1.4709 -0.3815  1.0262  0.0245  0.9941 -0.3572  0.3600 -0.8737
#>  0.3286  1.2483 -0.2192 -0.8237  1.2775 -1.0384  0.2957  0.2743 -0.9059 -0.1006
#> -1.4619 -0.0762 -0.1430 -1.4079 -0.7016 -1.1046  1.3046 -0.3124  2.3486 -1.0635
#>  0.9444  0.2973 -0.7965 -0.9571  0.4097  1.9955  0.4804  0.7742 -0.6487 -0.7245
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]