Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0222  0.9770  0.5049 -0.6348  0.0991 -0.1328 -0.3391 -1.3485 -1.8671 -0.3614
#> -1.3862 -1.4786 -1.0018 -0.3624  0.6296  0.8515  0.7025  0.0984  0.0471  2.1165
#> -0.4419 -0.3220 -0.7718  0.9044 -0.3492 -0.8977  0.4645  0.4868  0.6304 -0.0466
#>  0.1727  0.1873  0.1702  0.5516  0.7395 -0.4356  0.9325 -0.7535 -0.0592 -1.0266
#>  0.3815  0.3325  0.3597  0.7100 -0.1453  0.1279 -0.4716 -0.5150 -0.6887  0.5533
#>  0.2544  0.8681 -0.4140 -0.6374 -0.6798  0.6811 -0.5915 -0.4436  0.0714 -0.9321
#>  3.0953 -0.7177  0.8186 -0.3469  0.5052  0.1185  0.4556  0.8819  0.5998 -1.3630
#>  0.2174 -1.0224  1.1689 -0.9042  0.7968  1.2133  1.0731  0.3622 -1.7616  1.0596
#>  0.3944  0.0621 -0.6361 -0.5152  0.5932  0.9782 -0.8083  0.9807 -0.2459  1.9744
#>  1.0026  0.8157  0.0582  0.0304 -1.0230  1.0233  0.7051  1.3847  1.3997  0.2974
#> -0.8926 -0.4362  0.6650 -0.5213  0.8013  0.0373  1.1938 -1.9374  0.4919  0.5625
#> -0.0006  0.5211  0.3229 -0.2232  0.4174  0.1725 -0.7795 -0.7824  0.0688 -0.3083
#>  0.3686  0.1701  1.5741  0.1140  2.0249 -1.2881  0.4922  0.3680 -0.6618 -2.0442
#> -0.7345 -0.1969 -1.1961 -0.6725 -0.3665  0.8631  0.4565  0.9200  1.3815  0.0510
#> -2.2198  0.7542  0.9003  0.8114  1.4223  0.0093  0.8718 -0.6260  0.4197  0.4339
#> -1.1541  0.2321 -0.2727 -0.1004  1.1848  0.0681 -0.0164 -0.5454 -0.6555 -0.3025
#> -1.7133  0.0140  2.4813  0.3355 -0.7051 -0.4180 -1.8359  0.0004  1.6838 -0.5982
#> -1.6470  0.0437  0.4169  0.0896  0.1363 -0.0262  0.2637 -1.4906  0.4944 -0.8942
#>  0.1125 -0.8421 -0.1896 -0.7372  0.7981 -0.4231  0.9986 -0.6174 -0.7571  0.1122
#> -0.6614  0.1477 -0.4456 -1.7393  0.4596  0.1476  0.0023 -1.2479  0.5071 -1.8663
#> -0.6102  0.2278 -0.5626 -1.2532 -1.1327 -0.7740  2.0027 -0.6548  0.3945 -0.0763
#> -1.0104  1.3409 -0.7308 -0.8462 -0.4025 -0.3759 -0.0666  0.9744  0.3830 -0.8204
#> -0.4796  0.1803  2.2160 -1.0697 -0.1459  1.2800 -1.1489  0.8732  0.1937  1.4647
#> -1.0785  1.3479  1.3274 -1.1128 -0.9519  0.5199  0.2632 -0.9649 -0.3727  0.7898
#>  0.5276 -1.2594 -0.7744 -0.1728  1.9328  0.2056 -0.3163  0.1323  1.4229  0.6464
#>  1.2045 -0.0016  0.6792 -1.6119  0.2255  1.2877 -0.7598 -2.2702 -0.8104 -0.6466
#>  1.0121 -0.2449 -0.3942 -0.8348 -0.1192  1.1126  1.5544 -0.2511 -0.0055  1.6191
#> -0.6289  0.9619  0.7852  1.4309  1.1850  0.5856 -0.0359 -0.5511 -0.1355 -1.4758
#>  0.0872  0.7541  1.2351 -0.9881 -2.3701 -0.4766 -0.4059  0.1713  1.4204  0.5434
#>  1.5265  0.7010  1.5267 -0.4217  0.3050 -0.1329 -0.7450 -0.6341  0.7095 -0.7071
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]