Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-0.0745 -0.2811 -0.1844  0.2746 -2.0463  1.2823  1.2078  0.7694 -0.5278  0.0514
#>  0.7085 -0.8718 -0.0222  1.4160 -1.9735 -2.2388  1.7941 -1.3565  1.2263 -0.3119
#> -0.2198  0.0747 -1.5146  0.5784 -0.1417  1.1323 -1.1579  0.6230 -0.2892  0.6576
#>  0.4494 -0.3999 -0.8777  1.2917 -2.0992  0.8560 -0.1175 -1.9722  0.8048 -2.1523
#> -0.9893  0.3600  0.6532 -2.0273  0.9877 -0.4889 -1.4296  0.4619 -1.0729  2.1129
#>  0.2928 -0.3583  0.5822 -0.0259 -0.4146  1.0943  0.1337  1.2903  1.5154 -0.0415
#> -0.6364 -1.2122  1.2006 -0.4918  0.8743 -1.3768 -0.3521  0.6304  0.8946 -0.8392
#> -0.3164  0.6031  1.2423  0.1190  1.2090 -1.0200  0.7972 -0.9011  0.2466  0.0180
#>  0.0387  1.1171 -0.3689 -2.3314  0.4191 -0.8534  0.4444 -1.2531  0.6615 -0.8681
#> -0.5510 -0.1682  0.9798 -0.1404 -0.1353  0.2428 -0.2158  1.6503 -0.5358  0.9430
#> -0.0968  1.3338  1.6721 -1.8302 -0.7868  1.0105  1.4909 -0.0109  1.3243  0.8884
#> -0.0062  0.2625 -0.5544  1.2227  0.9371  1.8381 -0.0235  1.2622  1.0962  1.4478
#> -1.0606  0.4712 -0.7181 -0.5876  0.8337  0.2950  0.8670 -0.2119 -1.7049 -1.1302
#> -1.8453  0.9776 -0.0904 -1.1037  0.6817 -0.6996 -0.2998 -0.8964  0.4836 -0.0550
#> -0.0945  0.5104  0.6746  0.0937 -0.0521  0.1565  0.0766 -1.2789 -1.1712  0.7493
#> -0.2636  0.0135  0.4769  0.0579 -0.1604 -0.8477 -1.2429 -0.6474 -1.1664  1.0937
#> -1.6153  0.1039 -1.1938 -0.5650 -0.1256 -1.6734  0.6821  0.0085  0.0948  1.2767
#>  1.2125 -1.1456  0.1084 -0.3245 -0.3357 -0.0544  0.7428 -0.0771  0.3637 -1.1847
#> -0.5766 -0.3737 -0.0896 -0.0762  0.2651  1.5460  0.1659 -0.6351  1.6782  0.3535
#> -1.4815  0.6216  0.8107  1.3005 -0.1517 -0.3634 -0.6758 -1.7010 -0.5840  0.5628
#> -0.2563  0.7811  1.6749  0.1646 -0.4140  0.6457 -0.6969  1.4651  1.7693  0.1335
#> -1.4899 -1.0424  1.5397 -1.2724 -0.7540  1.0029 -0.3427 -0.2341  0.2341 -2.0807
#> -0.2835 -0.5717  1.1331 -0.5960 -2.0703  0.0452 -1.5640 -2.4138  0.2703  1.0882
#>  1.3205  1.8371  0.0202 -1.0308 -0.3218  2.0597  0.8083 -0.7036  0.1543 -0.1936
#> -0.1905 -1.7182  0.5266 -1.9356  0.0776 -0.8088 -0.4223  0.3365 -0.5302 -1.9125
#> -1.2780 -0.7420 -0.4277  0.2129  0.7050 -1.0120  0.9617  1.0854 -1.1892  1.3280
#> -0.7715  0.9947  1.2426  0.0899  1.3950  1.4988 -0.4040 -1.7499  1.8118 -0.8588
#>  1.0381 -0.3767 -0.1207  0.2902  0.1554 -0.8964 -0.7641 -1.1281 -0.9263 -0.5600
#> -0.5006 -0.5530 -1.9960 -0.0030  0.8852  1.1621  0.9400  0.9779 -0.7785 -0.0082
#>  1.9972  0.8231  0.8989 -0.3351  0.3784  0.2183 -1.9816 -0.8674  0.5725 -1.7665
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]