Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.0048  1.0544 -0.1921 -0.0778 -1.2046 -0.3927  0.4245 -2.2158  1.5810 -0.9304
#>  0.4777  1.4210  1.1237 -1.5006  0.2952 -0.8202 -0.8797 -0.6687  0.5476 -0.0141
#>  1.1660  1.1248 -0.2576 -1.3626 -0.2833 -0.3218 -1.2481 -0.1085  0.7872 -1.1302
#>  0.1884  0.8618 -0.9296  2.4654  0.2476 -1.2600  1.0684  2.5143  0.3644 -1.3442
#>  0.6328 -0.2673  0.1959  0.6835  0.1936 -1.0402  1.4539  0.2023 -0.8739  1.5457
#> -1.2480  1.2290  1.1386  3.1659 -1.4587  1.2031 -2.0179 -0.4889 -2.3379 -2.3200
#> -0.7688 -0.0844  0.4276  0.9091  0.8162  1.1754 -2.3672 -0.2487 -0.7005 -0.1596
#>  1.1224  0.9715 -1.1633  1.6560 -0.3249 -0.9174  0.1196  0.5977  0.6289  0.8475
#>  0.5265  0.7635 -1.0131 -0.0470  2.4143 -0.7619  0.0036  1.2791  1.3086 -1.1391
#> -0.3997 -1.8076  0.0700  0.3327  1.1525  0.8422 -0.0288 -1.3383 -1.4342 -1.9797
#>  0.2772  0.5646 -0.3844 -0.2605 -1.0708  1.2441 -1.0492  1.0700 -0.2510  1.1655
#> -1.7737 -0.7336 -0.0756 -0.3693  0.6017 -1.2264 -1.6853  1.7930 -0.2241 -0.8787
#>  0.4293  0.2419 -0.8429 -0.9070 -0.2542  1.1167  1.4889 -0.1780 -0.8110  1.5234
#>  0.3134  1.7447  1.3422 -0.0085 -1.2380  2.0475 -0.2000  0.7676  0.8139  1.1464
#>  0.0606 -0.2446  0.6442  0.3539 -0.1609 -0.5777 -0.1002  0.4410  0.2160 -0.2593
#>  0.5851 -0.0799  0.5778 -0.1236  0.3677  0.4095  1.0311 -0.2007 -2.5555 -1.9472
#>  0.0564  0.5964 -0.4898  0.8589  0.2186 -0.9834  0.9947 -0.5875  0.0781 -1.9891
#> -0.8720 -0.2379  0.4839  0.2492 -0.1890  1.3584  1.3402  0.8106 -0.9369  0.7179
#>  0.7754 -2.2824  1.7201  0.9186  0.4823 -1.4165  2.0552 -0.5218  0.0113  0.4249
#> -0.3443 -0.1217 -0.5397  1.2744 -0.8527 -1.0153 -0.1758  0.0064 -0.1950  0.0090
#>  0.5103  0.1759  2.3201  0.0440  0.5959 -0.4908  0.3578  0.2262  0.8966 -0.7755
#>  0.6269  0.5639 -0.3223  0.8377 -0.9102 -1.5405 -0.4397 -0.0090  0.0717  0.4629
#>  0.1275  0.2417  1.8155  1.6264 -0.8730 -2.0051  0.9335  1.6712  0.6218 -0.1856
#> -0.3158  0.3761 -0.1758 -0.2302 -1.8032  0.4182 -0.2469  0.0306 -0.0233  0.0927
#> -0.4316 -1.0347  1.2020 -0.7854  0.2829 -0.4650  0.7218 -1.3182  0.0425 -0.3113
#> -1.4585 -0.7942 -0.6878 -0.1098 -1.4524  2.3110  1.0814  0.3695 -0.2502  0.1379
#> -0.3167  1.3292  1.5770 -0.4002  1.1721  0.0135  1.0399  0.2628  0.4693  0.6432
#> -0.0403 -0.3176 -0.4617  1.2353 -0.7250 -1.3810 -0.2556 -0.8438 -0.3907 -0.7761
#> -1.3930 -1.3759  0.9675 -0.0940  1.5237  0.9617  1.2628  0.5755 -1.2262 -0.5744
#>  0.5050 -2.0393  0.2659  0.2644 -0.3771  1.2716  2.2560 -1.0861  0.2624  0.2062
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]