Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5273  0.0910  1.2516 -1.9514  1.3250 -0.0090 -0.1229 -0.1071  1.4460 -1.1707
#>  0.2159  1.4967 -0.7328 -0.8196  0.2275 -1.3806  0.0066 -0.1807 -0.4832  0.0782
#> -0.2676  0.9621 -0.1360 -1.2923  1.7529 -1.6581  1.5048  1.0138  1.2280 -0.6645
#>  0.0363 -0.0874  0.1391 -0.5497  2.0101 -0.1710 -0.4823 -0.2880  1.5035  0.2181
#> -0.7192 -0.0385  0.9196  0.2476  0.8472 -0.0413  0.6460 -0.8485  0.6094  0.7795
#> -0.3350 -0.5339  0.2461  0.6294 -0.4480 -1.0241  0.5971 -1.2722 -0.4573  0.5366
#> -0.6411  1.1438  0.6798  1.6639  1.0269 -0.9391 -2.0689 -1.3067  0.8343  0.6917
#> -0.4800  0.2327 -2.5212 -1.5390  0.9520 -0.4127  1.3510  0.7520 -0.3836 -1.0373
#> -1.0945 -0.9633  1.7206  0.8904 -0.5674  1.3426 -1.5444  1.0241 -1.1437 -1.9731
#> -1.6877 -0.7602  0.5391 -1.0643 -0.3676 -0.1917  0.5950  0.8802 -0.4474  0.9764
#>  1.2098 -1.2040 -1.0405 -0.7046  1.4313 -0.0525 -2.5434 -2.7508 -0.4700  1.4918
#> -1.5672 -0.1388 -1.0862  0.8074  1.4409  0.4976 -0.7375  1.1249 -1.6222  0.0808
#> -0.0185 -1.0310  2.2889 -0.5294 -0.6121 -1.9572 -1.5575  0.7062  1.6408  0.2396
#>  0.3831 -1.5180 -0.3196 -1.1771 -0.6728  0.5776 -1.0305  0.2063  0.6476 -2.4279
#>  1.3112  0.3468  0.3861 -1.4353 -1.1553 -0.9499  0.0398  0.6046 -0.4861  0.0220
#>  1.1545  0.7721 -0.6518  0.3412 -1.2204 -0.9846 -0.6740  0.9053 -0.8876 -1.0145
#>  0.6340 -0.6674 -0.4305 -0.1116  0.6885  0.7351 -0.8498  0.3561  0.6999 -0.9807
#>  0.7518 -0.7264 -0.9110  1.3120 -1.5321  0.6448 -0.3170  1.1205  0.2240  0.2881
#> -0.1520 -2.4492  0.7666 -0.6245 -0.5225  0.2997 -0.6379  0.1477 -0.5651  0.5820
#>  0.6561 -0.6238 -0.8216 -0.3066 -0.4089 -0.5338  0.6865  0.8823 -0.9984  0.9311
#>  0.4160  1.3462  1.7450 -1.6995  0.0998  1.5021  0.9719 -1.2107  1.6876 -0.5728
#>  0.9093  2.4286 -1.0140 -0.1725 -1.1260 -0.2393  0.6566  2.1120 -2.0861 -1.2015
#>  1.0912 -0.5376  0.2928 -0.2846  0.6481 -2.2632  1.1327 -1.5355 -1.7393 -0.7014
#>  2.0536  0.2459 -0.7082  1.4324 -0.1309 -0.2011  0.2373 -0.4914 -0.8658  1.4897
#> -0.9354 -0.9026  0.0455  0.0762 -0.1186  1.0244  0.4671  1.4699  1.2420 -0.3023
#>  0.8365 -0.0595 -0.6369 -0.4674  0.3762 -0.0801  0.3784 -1.9791  0.2879  1.4096
#>  0.3877 -1.2211  0.8549  2.0888  1.4894  0.4826 -1.0248  1.1353 -1.6522  2.4468
#>  1.7632 -0.0926  1.4182 -1.2310  0.1747 -0.2333 -0.4136 -1.1557 -0.3383  0.5252
#>  0.1699  0.4846  0.9788  0.6229 -0.3949 -0.1486 -0.6977  1.4142  0.2642 -0.3086
#>  0.0052  1.6071 -1.5806  1.3362 -1.5149  0.3173 -1.5439 -1.8993 -1.4057  0.4443
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]