Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.0226  0.9171 -0.7454  0.4447 -0.4255 -0.0343 -0.1250 -0.7107 -1.2656  1.8826
#> -1.4903 -0.7326  0.3337 -0.9962  0.3553 -0.7264  0.0692  3.0328 -1.0060 -0.6063
#> -1.5372 -1.4615  1.3523 -1.2083  1.0549 -1.7256  0.2895  0.0347 -1.0994  0.3067
#> -0.0663 -0.8078  0.7291 -0.3132  0.5676  0.6198  0.5867 -0.5572  0.2524 -1.9847
#>  1.0654 -1.5042 -0.1491  0.0570 -0.4158  0.9061  0.4125  1.5880 -0.8034  0.7454
#>  0.9120 -0.4485 -0.1671 -0.5527  1.2210  1.6078 -2.1194 -1.0891 -0.5743  0.3546
#> -0.1457 -0.8375  0.9691 -0.7844  0.8190 -1.4628  1.1655  0.0847  1.0109 -0.2427
#>  1.8003  0.9940 -1.2898 -0.8871 -0.0741  1.0152 -1.7865  1.2891 -1.7220 -0.1287
#>  0.7874 -2.0245  1.6103 -0.4614 -0.7158  0.1120  0.5591 -1.5182  0.4838 -1.0168
#> -1.9759  2.7077  1.6408  1.5428 -0.1474  0.2962  0.0628  0.1334  0.9557 -0.7608
#>  0.8030 -0.6957 -0.0710 -0.2115  1.9166 -0.7608 -0.1523  0.0671  0.1338 -0.5563
#> -1.2738  0.2474 -0.0240 -0.4595 -0.0450 -0.1135 -0.5320 -0.3328 -1.1244 -0.5290
#>  0.3638  0.5212 -0.8212  0.2562 -0.4176 -1.0956 -0.8068 -1.7550 -0.7385 -0.5170
#>  0.9571 -0.6986 -0.1799 -1.2618 -1.7705 -0.8375 -0.0224 -1.4785 -0.7540  1.2987
#> -0.7601  1.9342  0.0164 -0.7314  0.4730 -0.5218  0.7472 -0.3210 -1.4602 -0.3850
#>  1.8676  0.8672 -0.7622 -0.2321 -0.9083 -0.1673 -1.8831 -0.1050  0.9336 -0.2790
#> -0.3122  0.5714 -1.1751 -0.6542  1.3805 -1.2179 -1.1217 -0.9205 -0.3773  0.0326
#>  1.3034 -0.3486 -0.6065  0.3431 -0.9653  0.7773 -0.5056  2.0988  1.2492 -1.7170
#>  0.5334 -1.0037  0.3424 -1.2834  0.5073  0.6890  0.0520 -0.2200  1.1319  1.8557
#>  0.6389  1.0107 -0.7813  0.7609 -0.7253  2.4433  0.8817  0.6917 -0.9266 -1.2457
#> -0.9809  0.6934  0.9249 -1.1010  0.5795 -0.3402  0.2991  0.4307 -0.2984  1.3284
#>  1.1424  1.4786 -0.7796  0.0162 -0.5918  1.0906  0.1420 -0.2152  0.6957  1.8233
#>  0.8963  0.3811  0.0871  0.6547 -0.9080  0.4725 -0.1888  0.0192 -0.8536  0.4678
#> -0.4970  0.0862 -0.3171 -1.0262 -0.7677  0.8869 -1.3625  0.3837 -1.0520  2.5760
#> -0.0036 -0.0869 -1.0737 -1.0043  0.7227 -0.2545 -1.0203  0.7234 -0.4020  0.8605
#> -1.4350 -0.6089 -0.2324 -1.5945 -2.1496 -0.5207  0.3497  1.1664 -1.7175  2.8461
#>  1.0332  1.2653 -0.5608  0.6090  1.6714 -0.3261  0.6123  0.8639 -0.7357 -0.4412
#> -1.6872  0.4703  0.8140  1.0523 -1.2983 -1.7250 -1.0678  0.1328  0.1766 -0.3419
#>  0.4789 -0.5647  2.1728 -1.7585 -0.9171  0.3653  2.3539  1.0518  0.1833  2.9634
#>  0.4998  0.4110  0.4644  0.7001 -1.1758  0.4451 -1.0231 -0.0553  1.1415 -1.0215
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]