Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.6484 -0.1760  1.4269 -1.3465 -0.2787 -0.0642 -0.7884  1.4312 -0.5645 -1.1443
#> -2.3758  1.4992 -0.3585  0.3384 -1.5962 -0.1797 -0.0433 -0.0026 -1.8858 -0.2233
#>  0.4583  0.4497  1.7560  1.1464 -0.4848  0.7552 -1.2267 -0.0656 -0.2280 -0.1350
#>  0.8517 -0.6396 -0.5026 -0.2412 -1.0576 -1.2015  0.5105 -1.3754 -0.2208  0.3538
#>  0.5575 -0.3389 -0.1403 -1.3432  2.0998  0.5187 -0.1743  0.5511 -0.4189  0.9249
#> -0.3703 -0.2257 -0.2540 -2.0088 -0.4774  1.5103 -1.3703  0.8924  0.2601  0.1856
#> -1.2103  0.9112 -1.1775 -0.8572  0.5865 -1.1172 -1.7068  0.2079  1.4488 -0.5802
#>  0.6843 -0.7041  0.8568  1.7006  0.5226 -0.8264 -1.0212  1.0135 -0.2709  0.8425
#>  0.2991  1.6712  0.5999  0.5144  0.5835 -0.6560  0.5616  0.1290  1.5181  0.8703
#> -1.3002  0.8414 -0.4005  0.7375 -0.4280 -0.5201  0.4507 -0.5083  1.5671 -1.1073
#> -0.9915 -0.3461  0.4554  1.4340  0.1178  0.5377 -0.2170  0.1423 -0.7617  0.3363
#>  0.0489 -0.3749  0.9480  0.3977 -0.5852 -0.0833 -0.3737 -3.3891  1.5569 -0.7110
#>  0.8932 -1.4698 -0.1846 -1.6311 -2.0379 -0.1152  0.7046 -1.4263  0.1386 -1.4005
#> -0.0845  3.3180  0.3277  0.7649 -0.8891 -0.4151  0.9104 -1.2062  0.7984 -0.6551
#> -0.1322  1.1834  0.2052 -0.7449  0.2778 -0.3108 -0.6305  0.2526  0.9746 -0.1700
#>  0.4405  0.3542  0.8628 -0.9100  1.3257  0.0306 -0.8502  0.6663 -0.4081 -0.3156
#> -1.1611 -0.0200  0.6850  0.1682 -0.4804 -1.3421 -0.6145  0.2233 -0.0948  0.1971
#>  0.1858 -0.1000  0.2615  0.4601  0.0647 -0.0764  2.6794 -1.0511 -0.1050  0.4505
#> -0.0778  0.2642 -0.6090  1.6381  0.7071 -0.0950  0.0195  0.7091  0.4811  1.0885
#>  0.3417  0.1606  0.1217 -1.1722  1.1994  1.1171  1.4155  0.4938  0.7651 -0.2933
#> -2.1072 -1.1403 -1.0234  1.2948 -0.5149  1.1745  0.2931  0.6147 -1.9371 -0.2848
#>  0.1116  0.6780 -1.5066  0.1359 -1.0776 -0.4100  0.2924  0.5018  1.5896 -0.0456
#>  0.2936 -1.0005 -0.2308 -1.0880  0.1074 -0.5483  0.1842  0.6974  0.3471  0.1836
#>  1.0164  0.0681 -0.1571 -0.3034 -0.0631  0.7068 -0.1442  1.3780 -0.7055 -0.5442
#>  1.3067 -1.2992 -0.6596 -0.6207  0.0141 -0.8339 -1.2230  1.3668  1.3937 -0.0190
#>  0.4780  0.9540 -0.1171  0.2858  0.4015 -0.5217  1.0236 -0.7718  0.2670  0.4139
#>  1.2912  2.3400 -0.9861 -1.6210 -0.2518  2.0132 -0.6900  0.1594 -0.0127  1.4772
#> -1.0451  1.5248  1.6284  1.4642  0.0408 -1.1491  1.0166 -0.7725  0.5160 -0.0067
#>  1.2565  0.4653 -1.4443  0.2228  0.2123  0.1875 -1.3163 -0.1987  1.2943  1.8855
#>  0.1814 -0.4616 -1.7321 -0.9538 -1.3178 -0.6717 -0.3067 -0.9164  0.4337  1.7085
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]