Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.5917 -0.5382  0.4375  0.8504  0.7749 -0.1812 -0.5807 -0.4555 -0.4390 -0.5001
#> -1.1726 -1.7799  0.1855  0.6187 -0.7465 -0.7440 -0.5218  1.6725 -0.5745  0.2344
#>  1.7862  1.5341 -1.5453  1.3416 -1.6278 -0.1563 -0.6860  0.1770 -1.0616 -1.6079
#>  1.1490  0.3677  0.6088  1.0091 -0.4884 -0.5232 -0.5355 -0.2306  1.9293  0.0989
#>  0.0870 -2.1867  0.2207 -0.0978  0.9625  1.4062 -0.0444  0.5944  0.7294  1.0126
#>  0.8295  0.9492  1.7764  1.1453  0.2083 -0.5600  0.0164  0.5998  0.9398  0.6317
#> -1.3628  0.8981 -0.0823  0.1013 -0.3856 -0.0434 -0.7144 -1.4980 -0.6156  0.1788
#> -0.0927  0.2603 -1.2203 -0.1849 -0.8333 -0.3924 -0.9603 -0.5103 -2.3982 -0.2894
#>  0.7855  0.2467 -0.8091  0.2626 -2.1730 -1.0395 -0.3674  0.3945 -0.3185  1.0734
#>  0.3100 -0.9352  1.6675  1.5775  1.8514 -0.0597 -0.4857 -1.4166  0.3314 -0.8833
#> -0.6066  1.4495  0.5041  0.8425 -0.0482 -0.3732  0.3629 -0.1448  0.8652 -1.3373
#>  0.2486 -1.7848  1.1305 -0.3600  2.0800 -0.5437  0.4511  0.0363  0.4196 -0.6092
#>  1.8813 -0.5465 -0.3919  0.4417 -0.0278 -2.9836  0.3043 -1.0156 -0.1062 -0.0130
#>  0.5655  0.7370 -0.2628 -0.8763  0.2831 -0.9133  2.7753 -2.4410 -0.7831  0.4364
#> -1.3593  0.2858  0.8107 -1.3642  0.3698 -1.4036  0.5042 -0.9188 -2.2545  0.6618
#>  2.4696  0.5983 -0.7365 -0.4652 -1.2463 -0.4263  1.8288 -0.4989 -1.1738 -0.0499
#> -0.0653 -1.1926 -0.7715  0.5992  0.0972 -0.4946 -2.1748  0.2088 -0.5418  0.0240
#> -0.0101 -1.1186 -1.6071  1.1293  1.7119  0.1545 -0.5674  0.3461 -0.9359 -0.8560
#> -0.7527  0.7632 -0.7333 -0.8296  0.3563  0.3919  0.6185 -0.3656 -0.8961  0.2794
#> -1.0176  0.0117  0.1442  0.7815  0.2960  0.0533 -1.0013  0.4823 -1.1704  0.4075
#>  0.6753  1.6912  1.4306  0.8760 -0.3534  0.7022 -0.3978 -0.0078  1.5417 -0.5626
#> -0.9046 -0.2262 -1.8684  0.0070  0.9997 -0.2472  0.5986  0.1072 -1.8546  0.4496
#>  0.1684  0.7412  1.0374 -0.7524  0.9707  0.8416  0.4411 -2.2253 -0.3864  0.7803
#> -0.8115 -2.0571 -1.0292  1.8287  2.9065  0.6404 -0.1741  1.0799 -0.5526 -0.6047
#>  0.1965  0.8923  1.2159  0.6841  0.6494  1.0651 -0.9900  0.8222  0.1502 -0.7708
#> -0.6618  0.4278 -1.1418 -1.9567 -1.1873 -0.1562  0.0880  0.4533  0.6944 -1.7247
#> -0.6129  2.4315 -0.2239  0.4236 -1.3657  0.4606  1.5738 -0.3457  0.8196  0.2505
#>  1.7569  0.4935  0.2459  1.8987  0.5590 -0.2248 -1.4806  0.8753 -1.6435 -0.3598
#> -0.3437 -0.2876  0.3441 -0.2109  0.2476  1.2492  0.8772  0.0094 -1.8749 -0.0007
#> -0.5025  0.5502  0.1215  0.2522 -1.1771 -0.9927  2.4400 -1.3493 -1.1277  1.4780
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]