Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.0863 -0.7151  1.5753 -0.5550 -1.6990 -1.5486  1.1155  0.4856  0.4108 -0.4938
#> -0.3258 -0.9900 -1.1221  0.8566  0.0600 -0.5165 -0.1426  1.0053 -0.7466  1.2182
#> -0.0381 -0.5530 -0.1857  0.2461 -1.1078 -0.8503  0.2578  0.7608 -1.5833 -0.8496
#>  1.6713 -0.2043  0.1276  0.6284 -0.9466  0.0020 -0.1075  0.4504 -1.1931 -1.3094
#> -1.7746 -0.0602  0.2213 -0.6690  0.4705  0.2703 -1.2449 -0.1339 -0.1287  0.7676
#>  0.3008 -0.6968 -1.7737  0.3583 -1.0024 -0.5357 -1.2680  0.3515  0.4172  0.6546
#> -1.2956  0.6275  1.7665  1.0303  0.5490  0.0788  1.1245  0.3066  0.1599 -2.1063
#>  1.0240  1.3930  1.3098 -1.5246  1.5162 -0.4225  0.4522 -1.3145  1.6858 -0.4862
#>  0.1578 -0.9894 -0.8936 -0.6000  0.7683 -1.8762  1.5696 -0.5435  0.0045  2.6988
#>  1.0408 -0.3626  1.1849  0.0881 -1.2819  0.3316  2.2550 -0.3259  0.1370  1.1428
#> -0.4414 -0.2924 -0.3123  0.6832 -0.4170  1.9048 -0.6398  1.1151  0.7963  0.7441
#> -1.2059 -0.0908  0.9600  0.5697  1.4936  0.7848 -0.5148  1.1966  0.4062 -0.0727
#> -0.7814  1.8790  1.6819 -0.6106 -0.0968 -1.1851  0.2318  0.2085 -0.6021  1.5670
#>  0.1335 -1.0440 -1.5343  1.1995  1.3079 -1.2424 -0.1718  1.3475  0.0554 -1.3447
#> -1.0885 -1.1512 -1.3189  1.3063 -1.1674  0.5521  0.3822  0.2180 -0.9104 -0.7444
#>  1.3282 -0.0747  0.5058  0.2898 -0.8250  1.9196 -0.2549  0.3681 -0.7165 -0.7629
#> -1.9555  1.1990 -0.1907  0.8603  0.6557  0.9524 -0.8207  0.3353 -1.2653 -1.0698
#> -0.8247 -0.2584 -1.1643  0.0811 -1.6318  1.1746  1.9672 -0.3253  0.3728 -0.2376
#> -1.8300  1.9665 -1.0831  1.1064  0.9060  0.2862 -1.2670  1.9372 -2.4232  1.2555
#>  0.0127  0.4653 -0.2403  0.8544 -0.2304  0.7390  0.7216  0.8965  0.1401 -0.6501
#>  0.3921  0.5285 -1.2816 -0.1096 -1.1900 -0.0957 -0.6875 -0.8061  1.4036 -1.6582
#>  0.6027 -1.3286  1.1319 -1.1409  0.2633 -1.0785  0.5627 -0.0538 -0.1926  0.5827
#>  1.1069  0.0868  0.0053 -0.0394  0.8669 -0.1904 -0.3601  0.6674 -1.3947  0.1146
#> -1.1673 -0.5039  0.8319 -0.4078 -0.6648 -0.5283  0.1537  1.9522  0.3885  0.9617
#> -0.1277 -0.5525 -0.1853 -0.9316 -0.5910 -0.2563  2.5578  0.9148  0.7823  0.0850
#> -0.0604  0.2993  0.7876 -0.2775 -0.8516 -0.0441  0.3101  0.5526  0.1244  0.0901
#>  1.7232 -0.5939 -1.9470  0.7640 -0.3347 -0.1614 -0.7084  0.0426 -1.0163  0.1506
#> -0.2301 -1.3713 -2.5335 -1.1119 -1.0092  0.2195  0.7046 -0.6646  0.9435 -0.7624
#> -1.7983  0.2470  1.3740 -0.3700  0.2772  0.6287 -0.5479  0.7556 -0.3497 -0.7446
#>  1.6462  1.1671  1.5655 -0.1174 -1.7227  0.1709  0.3177 -0.9012  1.0832 -0.2117
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]