Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)

}
#> torch_tensor
#> Columns 1 to 10-2.0448 -2.0215  1.1291  0.1161  0.3704  2.0336 -0.4316 -1.8440  0.5652  0.0955
#>  1.8121 -2.0756  0.3269  0.5870  0.4452 -0.1216  1.5075  0.5685  0.2889  0.6207
#>  0.6648 -0.8124  0.2595  0.9963  0.5271 -0.2775 -0.9670 -0.1767  1.2377  0.9016
#>  0.3432 -0.4074 -0.6451 -2.2372  1.5578  1.2055  0.1490 -0.4198  1.4480 -0.7924
#>  0.2768  0.9162 -0.0927  0.5403 -0.0716  0.6694  0.3307 -0.0276 -0.2270  1.8575
#> -0.2580 -2.6968 -0.4002  0.9940 -0.7198  0.8966 -1.1699  0.2014 -0.3620  0.0495
#>  0.3638 -0.8451  0.4503 -1.4276  1.1323  0.3583 -0.3411  0.1872  0.2753 -1.0715
#> -0.8129 -0.0407 -0.2049  1.2514  0.6137  0.4031  0.5538 -0.0401 -1.2202  0.5564
#> -1.3693 -0.0313 -1.2635  0.5423 -1.1514  0.3475 -0.2260 -0.9857  0.9699 -1.6945
#> -1.8106  1.3070  1.2309 -0.2163  1.7185  0.6860 -0.7654  0.0229  1.4717  0.9178
#>  2.5356 -1.0164  0.9552 -1.2347  0.8780  0.6805 -2.0270  1.0338 -1.6052  1.3947
#> -1.1463  1.8494 -2.5147  2.8479  0.5365 -0.1198 -1.3139  0.6790  0.0103 -0.3284
#>  0.0380 -1.0147  0.5112 -2.4638 -1.2372 -0.3319  0.6462 -0.3885 -0.2859  1.3969
#>  1.7221 -0.4561 -1.0401  1.8850 -0.2215 -0.7577  1.3819  0.8010 -0.9474 -1.3494
#>  0.5606 -0.2119  0.9987  2.7750  0.9700  0.6338  0.2446  0.5893  2.1548  0.1123
#>  0.6892  0.4398 -0.0489  1.7857 -0.0512  0.3870 -1.4670 -0.1607 -0.5207 -0.4855
#>  0.0631 -0.9840  0.6669  2.3789  0.5290 -2.3577 -0.8069  0.1819 -0.5647 -0.0118
#> -0.7691  0.7437  1.3213 -0.5545 -0.2755  0.5766 -2.1875  0.9083  1.8789  0.6441
#>  0.9197 -2.1354 -0.5045  0.6574  0.3320  0.6496 -0.8214 -0.4965 -0.4499 -1.1978
#>  0.8705 -0.1432 -0.5331 -0.2318  1.2938  0.2860  0.4085  0.4630 -0.3853  0.3812
#>  0.1409 -0.2636  0.4754 -0.1469  1.7073  0.8183 -1.5366  0.7530 -0.2605  0.8123
#> -1.9496 -1.0613 -1.1399  0.6495  1.6541  0.0250 -0.4057 -0.8701 -1.0767 -0.2872
#>  1.2691 -1.7748  1.2705  0.8773  0.4382  1.3045 -1.2414  1.3452  1.4725 -0.9097
#> -0.8938  0.6545  1.3086 -0.2678  2.3169  0.1283  0.8021 -0.4806  2.0177  0.2372
#> -0.5466 -0.0952 -0.6784  0.8167 -0.1509 -0.3892 -0.1442 -1.1173  0.0018 -0.6459
#>  0.4989  2.2073 -0.4553  1.5376  0.0884 -0.3899 -0.8927  1.6839  2.4613 -0.8294
#> -0.8161  0.9731  0.6812  0.4232 -0.9305  1.4995  0.5708  1.0356 -0.9446 -0.1973
#>  0.8025 -0.1717 -0.3721 -0.1648 -0.7146 -0.3842 -1.3122 -0.4274  0.2944  0.8046
#> -1.1867 -2.6706  0.7564 -0.7882  0.4654 -2.1284  0.0932 -0.6409  0.1345 -0.0046
#> -1.2411 -0.6744  0.1336 -1.1706  0.4754  0.6950  0.1372 -0.5759  1.7660  0.6209
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]