Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.2698 -0.6213 -2.1309  0.7613 -0.8461  0.1961  0.7743 -0.5820  0.7491 -1.5683
#>  0.6656 -0.0391 -0.4236  2.7328  0.7148 -0.0365 -0.2337 -0.2663  0.8782 -0.0259
#> -0.7459 -0.6175 -0.4795  0.4947 -0.3440  0.0065  0.8748 -0.3308  1.7348  0.9898
#>  0.2084 -1.4515 -0.4195 -0.7289  0.4129  1.9615 -0.2830  2.0939  1.6982 -0.4835
#>  0.0881 -1.9295 -0.2978  0.4773 -0.3383 -0.0852  1.0951 -0.4279  0.0560  1.1989
#>  0.3290 -0.4386  0.2990  0.7432  0.0659  0.2535  1.3267 -0.1953 -0.5747  0.9472
#> -0.8889  0.8998  0.3152  0.4999  1.2286 -0.1590 -0.8689 -1.9080 -0.2211 -0.2919
#> -1.2184  0.7191 -0.4824  1.3979 -0.3732  0.0909 -0.3344 -0.3981 -0.9767 -1.5117
#> -0.1113  0.1244 -0.2803  0.8806 -0.2422  0.3774 -0.2376 -2.1335 -0.6625  1.3961
#> -0.2560  1.2350 -1.1243 -1.3070 -0.2731 -0.7668  0.9679 -0.0866  1.0746 -2.9947
#>  1.0088 -0.1196 -0.5697 -0.5395 -0.4811 -1.4640  1.0222 -0.8067  0.9587 -0.6016
#>  0.6364  0.1437 -0.6954 -1.3861 -0.5585 -0.5365 -0.7265 -1.1864  1.5389 -1.2534
#>  0.9005  0.2583 -0.4901 -0.7502 -1.0953 -0.1295 -1.6272  0.7359 -1.7421 -1.9220
#>  0.6889 -1.0601  0.9080  1.3742  1.3780 -0.1976  1.0857 -1.0652  0.4785  1.5549
#>  0.3771  1.4881  1.0683 -0.7746 -0.5718  0.9788  0.2873  0.1159 -0.6915  1.9577
#> -0.6339 -0.0369  1.1322 -0.3652  0.0806  0.4643 -0.3176  1.7282 -1.8592 -1.8093
#> -0.9614 -1.0851  0.4423 -0.4184 -0.2571  1.6120  0.8135 -3.0071  0.1311 -0.1699
#> -0.8651  0.3978  0.6795  0.0546 -0.7153 -0.8131 -1.7843 -0.1860  0.9388 -0.0992
#> -0.2253 -0.8181 -1.0222 -0.0009 -0.5740 -0.1999  2.5107  0.1499 -0.3666  1.5263
#>  0.1334  2.4621  1.2492 -0.0072 -0.7016 -0.7992  2.2364 -0.5387  0.1018 -2.6505
#> -0.5634  0.9956  1.3431 -1.6105 -1.3953  1.1309  1.7159 -0.1222  0.7888  0.6039
#> -0.5570 -0.3307 -0.5290 -1.2706  1.3370 -0.4143  1.7678  0.7209 -1.4120  0.1340
#>  1.1920  1.9582 -0.6612 -0.0424  0.8604  0.2077 -1.2626  0.4101 -0.7293 -0.4434
#>  1.0744 -1.4955  0.6311 -0.5256  0.9172  0.8315  0.5465 -0.3513  1.0425 -0.4102
#> -1.1622 -0.2835  1.1129  0.9183  0.3064 -0.0035  1.5714  1.9039  0.7625 -1.9366
#>  1.1119 -0.0660 -0.6391 -0.1358 -1.1638 -1.8504 -0.9830 -0.3717  0.4367  0.8589
#> -1.6907  0.2512 -1.3241 -0.1323 -0.8036 -1.0834  1.1062  0.1925  2.7662  1.2361
#>  0.2124  0.9088  0.7807 -1.3944 -0.4835 -1.8155 -0.4098 -0.6779  0.8322  1.0792
#> -0.7488  0.4549 -0.0202  0.7554 -1.0935 -0.9982  0.7972 -0.7603  0.2780  0.1193
#>  1.7429 -0.2441 -1.0277  0.7072 -0.4048 -0.8430 -0.2935  0.8071  0.8873 -0.1297
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]