Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.1098  1.7417 -0.5977  0.5468  1.4292  0.2381  1.2421 -0.7273 -1.1310 -0.7907
#> -0.7409 -0.8896  0.9529 -0.5602  1.7497  0.4811 -0.5345 -1.4343 -0.0812 -0.5573
#>  0.2061  0.6557  1.2442  0.1966  0.7653 -0.4700 -0.7987 -0.8247 -0.1134 -0.0506
#>  1.0600 -0.3895  0.6548  0.1669 -0.1913  1.1735 -0.3680  0.6439  0.5697 -0.7078
#> -1.0117 -0.1445 -0.8111  0.9937  0.5714 -1.1452  0.6335 -1.0450  1.8429  2.0141
#>  0.9019  0.7681  1.3721  0.2321  0.2820 -1.4963 -0.1879  0.2982 -0.3715  0.3416
#> -1.0144  1.2441 -1.3847  0.0213  1.6583 -1.0263  0.9959  0.6549 -1.1462 -0.1706
#> -0.2266  1.0177 -1.3686 -0.6281 -0.5479 -1.0276 -0.6892  1.1023 -0.4189  3.2675
#>  0.1167  0.6854 -0.2811  1.0282  0.5764 -0.6555 -0.0201  0.1040  0.4652  1.1687
#>  0.3644 -1.1089  0.4739  0.8617  1.0050  1.3367  0.9903 -0.7829 -1.8736  0.7815
#>  0.1489 -0.4126  0.4673  0.2077 -1.1144 -0.3506  0.3251  0.5515  0.6129 -0.7172
#> -0.0672 -0.3203  0.4649  1.0456  0.6377 -0.7412  0.1059  1.5598  2.1153 -0.7993
#>  0.9521  0.6103 -0.0657  0.1947 -0.8377 -0.6934  0.3659 -0.4237  1.0795 -0.4574
#> -0.5330  0.1759  0.0263 -0.2414  0.0931  0.9984  1.2447 -2.0085  1.0800  0.7785
#>  0.0689 -0.9801  0.4441 -0.0072 -0.3766  1.3895  0.4143 -0.0398  0.4350  0.5772
#>  1.5368  1.5521  1.1526 -0.7782 -0.1556  0.5946 -2.0075 -1.9790  1.1611 -0.1081
#> -0.5309 -1.1630 -0.7469  1.2041 -0.4442 -1.1802 -0.4782  1.2801 -0.8300 -0.2062
#>  0.8472 -0.8478 -0.8145 -0.8561 -1.3476 -0.3964  1.6147  0.3549  1.4031 -1.6533
#>  1.2875  1.0318  1.3827  2.2716 -0.0564  1.9050 -0.5766  0.4957 -0.7578 -0.4985
#>  1.1421  0.3238 -0.2387  0.0690  1.5986  0.5888  3.8565 -0.7402 -0.2643  1.5790
#> -1.0838 -1.1487 -0.5253  1.3754 -0.4709 -0.0772  1.6824 -0.6946 -1.4093  0.3872
#>  2.2954  0.7533  3.0372 -0.0569  0.0780  0.5087 -0.2349 -0.7414 -1.7725 -1.8305
#> -1.4114  1.9166  0.5699  0.0056  0.7501 -1.1841 -0.8214 -0.9753  0.6056  0.5657
#>  0.5115 -0.7699 -0.0865 -0.8766  0.0723 -0.5133  0.3948  0.6813 -0.7323 -0.0791
#>  0.2687  0.7317 -0.1021 -0.8178 -0.4991  0.6283  0.3523  0.8282  0.0247  1.8525
#> -1.9651 -1.1398 -2.5318  1.0541  0.1256  0.8193 -0.2611 -1.3553  0.0222 -0.1410
#>  0.0699 -1.7552  0.6084  0.3527  0.5456 -0.1263  0.3377 -0.6769  0.7649 -0.6004
#> -1.2075  0.8877 -0.6301  1.5920 -0.4112  1.6507  0.0360 -0.9499  0.1722  0.0597
#> -0.1225  0.2838 -0.6632 -0.5747  1.3393  0.1258  0.1128 -1.6109 -1.5812 -1.0102
#>  1.1110  0.1765  1.0333 -1.4073  1.8580  0.6718  0.4549 -0.5615 -1.7464 -1.6401
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]