Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.2012 -0.0251 -1.6799  0.0847 -0.0612  0.4409 -1.2029  0.8603  0.6665 -0.6731
#>  0.0603  2.0559  0.5261 -1.0384 -0.6540  0.2309 -0.3916 -1.0176  1.0795 -0.0747
#> -1.1301 -1.2564  0.3425  0.9343  0.7416  0.3302  1.1168 -1.8006 -0.5641  0.0737
#> -0.7590 -0.6244 -0.6544 -0.7366 -0.3145  1.5654  1.5253 -0.4437 -0.1054  1.1976
#>  0.6597  1.3269 -0.8358  0.4201  2.0391 -0.2395 -0.6484  3.4588  0.9036 -0.0487
#> -0.4143 -1.3462  0.7123  0.2975 -0.0186  1.9889  0.6954  0.7958 -0.1650 -0.2044
#> -0.6309 -0.7119  0.5991  0.0015 -1.1655 -0.9518  1.2359  0.5029 -0.8868  0.6335
#> -0.0465 -0.5600 -1.8723  0.3054 -1.7897  1.3798 -0.3837  0.8707  0.0606  0.8981
#> -0.0188 -1.4251  0.5727 -0.5546 -1.9147  0.1095 -1.3426  0.9509 -0.5895 -0.5749
#> -2.2794 -0.0491  0.6641 -0.7041 -0.8341 -1.3971  1.8184 -0.4625 -0.5244 -0.8403
#>  0.0422 -0.0436  0.8680 -1.1402  0.4291  0.3471  1.3206  1.3575  0.2030  0.1315
#>  0.6618  0.9112 -0.1207  0.0903  0.0146  1.0149 -1.5953  1.8632  0.1819  0.9346
#>  1.5270 -1.7766  0.5203 -0.2009  1.5471 -0.6705  0.3800  0.9367 -0.4534 -0.2449
#> -1.4331  2.0941  0.9646  0.3214 -1.0029  0.6229 -1.2132  0.0343  1.5206 -0.3171
#> -0.6700  0.5032 -1.9267  1.5483  0.1043  1.1386 -0.3125 -1.4283 -0.3006  0.7640
#>  0.7521  0.5437  0.4068 -0.0296 -2.0627 -0.9231  1.8024  1.3369 -0.7992 -0.8222
#> -1.0642 -0.1087 -0.7815 -2.0904  0.8849 -1.8163  1.1444 -0.0558  0.5138 -1.0566
#>  0.6313  0.4309  0.4555 -0.2241 -1.5596 -0.5591 -1.0999 -0.4131 -1.3126 -0.3368
#>  0.1149  0.1657 -0.8158 -0.5963 -0.8980 -0.7998 -1.9223  0.8214  0.0552  1.1291
#> -0.7128 -1.1531  1.3907  0.7366 -0.6738 -0.6601  0.9085 -0.1807  0.5530  1.3515
#>  0.3124 -0.3440 -0.3182 -0.0865  0.1269  0.4863  0.8933  0.5631 -0.5710 -0.3497
#> -0.7564 -1.3431 -1.0608 -0.7021 -0.9200  0.0307  0.5671  0.9521  1.0489  1.3448
#> -0.1340  0.6606 -0.8749 -0.3899  0.5082 -0.1155 -0.2867  0.1364 -0.3318  1.2279
#> -0.3793 -1.1439  0.2360 -0.3945  0.7238  1.1269 -0.5408  0.3652  1.2669  0.0398
#> -0.1492 -0.4921 -0.5551  0.2921  0.5107 -0.3190 -0.2764 -1.9360 -0.6784  1.4859
#> -0.2089 -0.2389  0.1470 -0.3929  1.4968  1.8492  0.6869 -1.1281  1.9620 -0.2838
#>  0.4884  0.3212 -0.2656 -0.6345  0.7412  1.1004 -0.0366  1.0053 -1.6872 -0.1274
#> -1.0653  1.9425 -1.3592  0.5003 -0.9511 -0.9981  0.9937 -0.9986  1.7136 -0.6789
#>  0.5691 -0.6717  0.5402 -0.0551 -1.9269  1.0107  0.2161 -1.6606  0.6879 -0.1718
#> -0.5781 -0.2393  0.4633 -0.0945  0.6886 -0.5397  0.5474 -1.7636  0.8991  0.6731
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]