Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4920  1.1047  0.7436  1.5445  1.7225  0.6828  0.1167  1.1876 -0.8612  0.9384
#>  0.2563  0.7318 -0.4793  0.4450  0.4063  1.3493  0.3738  1.3401 -0.5484  1.8861
#>  0.1091  0.3070 -2.3414 -0.6591 -0.6448  0.5221  1.3134 -0.5959 -0.9223 -0.3964
#>  0.2220 -0.2002 -0.4683  1.0209  1.0129 -0.9701 -1.2320 -0.9467 -1.3198  0.8818
#>  2.7399  1.2117  0.2773  0.6620 -0.0618 -0.6281  0.8658  1.0338  0.3021  0.8746
#>  0.4945 -0.3690  0.5218  0.3764  1.2039 -0.8359  0.7198  1.1037 -0.4410  0.1627
#>  0.4902  1.6520 -0.2917 -0.2782 -0.5684  1.4114  1.1874  0.3997 -1.1677  0.2841
#>  0.2769  0.6841 -1.0158  1.8883 -1.3638  0.5087  0.4434 -0.4361  1.9102  0.1980
#> -1.9004 -0.5908  0.3613 -0.4996  0.6727 -0.4109  1.2364 -0.3086 -0.1205  0.7094
#> -0.6620 -0.7695 -0.5023  0.0212  0.5653 -1.1878 -1.4942 -0.5384 -0.3210 -1.4431
#> -0.3209  1.0281 -1.4061  1.4574 -0.9504  0.6804 -1.5564  2.0983  0.0959  1.4445
#>  1.2341  1.0076 -0.3623  1.7526 -0.6489  3.0822 -0.5147 -0.6682  0.5044 -0.8829
#> -0.1722  0.3803 -0.4662 -1.1543  0.4970 -1.4620 -1.5374 -0.5231 -0.2794 -0.1797
#> -0.9107 -2.8827 -1.6687 -0.1521  0.2189 -0.2217  0.8018 -0.9251 -0.0706 -0.9862
#> -1.2417 -0.6562  0.0015  1.1170  0.0220 -0.9187 -0.2974  0.3640 -0.0688 -0.0119
#> -0.6421 -0.9475 -0.6328  0.2153  1.3620 -0.3914 -1.2586 -0.2563  0.5883 -1.1850
#>  1.5616  1.1969 -0.9724 -0.5635 -0.7486 -1.0232 -1.9552 -1.5404  0.0101  0.3213
#> -1.0019 -0.4301  0.3585 -1.4615  1.6026 -0.6272  0.1572 -0.5805 -0.3600  0.0186
#> -1.3665 -0.1556  1.4485  0.0222  1.9405  0.7041  0.9891  0.0089  1.5507  0.1473
#> -0.1023  0.5883  0.0104 -0.3154 -1.3001  0.0624  0.1633 -3.6718 -0.1130 -1.5715
#>  0.1650 -1.7749  0.6439 -0.7069  0.7148 -0.0719  0.1161  0.7685  0.8753  0.6374
#> -0.4999 -0.3596  0.6023 -1.2441 -0.2739 -0.4525 -1.7926 -0.1425 -0.4639  0.2682
#> -0.4971  0.4190  1.7577  0.7299 -0.5546  0.1721  2.0283 -0.8631 -1.0012  0.3166
#> -0.8791  0.8576 -0.1599  1.0995 -0.4873 -0.6599  0.0857 -0.5576 -1.6816 -0.1436
#>  1.7049  1.0447  0.4864 -1.4328  0.4597 -0.3139  0.1035  0.1807  0.3409 -1.5609
#>  0.5213  1.8768 -0.5626 -2.0052 -0.5475 -1.2421 -0.6441  2.2860  0.3943  0.8273
#>  0.1547  0.4873 -0.5560 -1.1964 -0.5501  2.1713 -1.0431  0.0772  0.9487  0.5198
#>  0.7210  1.4525  0.1467  0.1918 -0.0536 -0.4459  0.1901 -0.2424  0.9358  1.0651
#> -0.2046 -1.2951 -0.6561 -1.0814  0.1108  0.3093  0.2751 -0.2437  1.1100 -0.2762
#> -0.3562  0.3269 -0.7354 -0.1327 -0.7426  0.8196  0.3722  0.5751  0.4244 -0.7808
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]