Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.1672  1.1400  0.4995 -1.7640  0.4462 -1.2417 -0.3235  0.0715  0.5639  0.6712
#>  0.2730  0.3841  0.9610 -0.2224  0.8652  0.0665 -0.1585 -0.0440 -0.5252  0.7836
#> -0.9877 -0.5288 -0.7437 -0.0252 -0.7646 -0.1483  0.0345 -1.2505  0.3061 -1.1524
#>  1.2580 -2.9444  0.0854  0.5235 -0.0235 -1.0317 -1.1919 -0.2841 -0.2426  1.5320
#> -0.3783  0.8104 -1.1120  0.3293  1.2687  0.7058 -0.6016  0.6547 -0.3631  0.6918
#> -1.3467  0.1511  0.9719 -0.1275 -1.0048 -0.6981 -0.6418  0.4111  0.1359  1.6472
#> -0.0864  0.9515  2.4505 -0.4431  0.2216  0.5439 -1.7737 -1.3443 -0.9054 -1.2245
#>  0.7279 -0.3949  0.8667  0.3753 -2.1031 -0.9667 -1.1147 -1.2428 -0.3817 -1.3367
#>  0.1128 -1.5013 -1.1392 -0.6298  0.0155 -0.5811  1.1950 -0.2346 -0.3180  0.7970
#>  0.5119  0.9085  0.1630 -0.6832 -1.0936 -0.4421  1.3373 -0.5059 -0.4516 -0.9512
#> -1.1807 -0.4389 -0.4085 -1.5640  0.3485  1.3064 -0.4032  0.2437  0.6501 -0.1657
#>  0.2758  0.0859 -0.8110 -0.0864  1.5669  0.4259  0.1055 -1.0132  0.2111  0.1306
#> -0.6264 -0.5076 -0.4793  0.6483  0.6395  0.0453 -0.2774  0.9658  0.6718  0.7102
#> -0.3138 -1.5481  1.0954 -0.5751  1.0847  0.0727 -0.8323 -0.1062 -0.4391 -0.8235
#>  0.9142 -0.1245  0.9158 -2.0568 -1.0025  0.5953 -0.6498 -0.2356 -0.5334 -0.5311
#> -0.8960  1.0279  0.2760 -0.3184  0.1841  0.8877  0.8127  0.2919 -0.3128 -1.0654
#>  0.7646 -0.1867  0.0540 -1.7017  0.8257 -0.0695  0.5577  1.3036 -0.6312 -1.8276
#> -1.3081  0.9730 -0.8631 -0.0534 -1.4703  0.0131 -1.3055  0.7723 -0.9594  0.0912
#>  0.5351 -1.3626  0.5690  0.1480  0.4208  1.2973  0.7737  0.2811 -0.8082  0.9754
#>  1.3107 -1.4199 -0.1340  0.4995 -0.5538 -1.2442 -0.2245 -2.0487  0.1071 -0.3414
#>  0.3884 -0.3087  0.7214  0.2696  0.2774 -2.4909 -1.2016 -0.6651  1.8764  0.9310
#>  1.2445 -0.8677 -0.6986 -0.0616  1.0503 -0.1780  0.2244 -0.7904 -0.1468 -0.6954
#>  0.0376  0.0982 -0.2130 -2.4195  1.0899  2.6013 -0.3154 -0.0944 -1.4338 -0.2902
#>  1.2995  0.0046 -0.7886  1.7954  0.1558 -0.1666  1.1498  0.1245  2.0557 -0.9414
#> -0.6851 -0.8813  0.5358  0.5132 -0.2331  0.2817 -0.1018  0.3439 -1.0410  0.6764
#> -0.6433  0.7179 -2.0695 -0.4384  1.6301  0.5619 -0.7277 -0.7198  0.5448 -0.3687
#> -1.0259 -1.0168  0.2513  0.0748  2.7994  1.1822  1.7962 -0.7775 -0.6565 -0.2263
#> -0.9992  0.5133 -0.1844  0.0166  0.1632 -0.0933 -0.5239  1.5080  0.8402 -0.5579
#>  0.4617 -0.2693 -0.4103 -0.5874 -0.9798 -1.7066  0.9642  0.0864  1.3010 -0.7260
#>  1.9660  2.2222  1.0981 -0.6005 -0.8863 -0.6926  1.7255 -0.6462 -0.0025  0.7005
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]