Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 1.8146  0.7422 -0.8458 -0.1789 -1.1315  0.6001  0.4717  0.2325  0.6159  2.4002 2.0874 -1.6125  0.5031  0.4299  1.5799  1.1990 -0.2562  0.1252  1.3778 -0.6159 1.4998  0.6829  0.7118 -0.5208  0.5666 -0.4583 -2.1264  0.1023  1.1038 -0.1223-0.5617  1.8613 -0.4074  0.9298 -0.1574  1.2543  0.3872  1.2228  1.3782  2.0427-0.8884 -0.4041 -0.5937 -0.8403 -0.8273 -1.0409  0.8948  0.2562  0.2336 -0.2257 0.4295  1.1108  1.6672 -0.0803 -0.0068  1.2713 -1.3651  2.0318  1.5271 -1.2057-0.5603  0.7336  0.9353 -1.2015  0.0180  0.1036 -1.1256 -0.4109  0.7216  0.8109-0.5746  0.9499 -1.3858  0.2301  0.5709 -1.4533 -0.0471  0.8337  0.7300 -1.9188 0.4379  0.2744  1.3709 -0.4624 -1.6003  0.6969  0.3164 -0.1534  0.5991 -0.6626-0.9750 -0.7081  0.3017  0.6998 -0.5610  1.8435 -1.3670  1.8504 -2.2035  2.3530 0.4511  0.6412  0.9897 -0.9088 -0.1734 -0.3652  1.8254  0.2249  1.4279 -0.4553 0.3816  1.9052 -0.7497  1.0045 -0.2626  0.1155  0.3253 -0.6147 -0.2404  0.0337-0.2138  0.2758  1.3122 -1.3350  1.4308  0.1168 -0.6621  0.8042  0.1644 -0.7972-0.3121 -0.3308 -0.0550 -1.8838  1.0999  0.2484  0.4109  1.1657 -0.1198  0.5383 0.6179  0.4794 -1.1946  0.9198  1.2249 -0.2475  0.9545 -1.2431 -1.1834  0.0507-0.1742 -0.4357 -0.2623  0.1330  0.1849 -0.3865  0.4323 -0.2525 -0.7301 -1.2665 0.9765 -1.6539 -0.0397 -0.0646  1.4096  0.0635 -0.5788 -1.0883  0.0922 -0.3994 0.7392 -1.5338 -0.0860 -1.7088  1.0429  0.9239  0.2496 -1.2321  1.3810 -0.5011-1.2569 -1.0849  0.1277 -0.3978  1.2207  0.7930  0.4381 -0.3201  0.6769  1.0726 0.4931 -0.4456  2.0466  0.0523 -0.6638  0.7455 -0.7155 -0.8893  0.8621  0.3463 1.6986  0.1493  1.0455  1.1558 -0.4327  0.5514 -1.6589  1.6116 -0.0106 -1.1591 2.1136 -0.0001 -1.4509 -1.6195  0.0334  0.0612  1.1560  0.0253  1.3540 -2.0595-0.4342  0.9041  0.5596 -1.6780  1.0783 -2.8806 -0.8661 -0.0157 -1.7751 -1.2504-0.7985 -0.0392 -0.0171 -2.0761  0.1173 -0.1355  0.4095  0.6726  0.1400 -0.4905-0.5316 -1.2731 -0.3515  1.2967 -0.8741 -0.2432 -1.5810 -1.5480  0.6829 -0.5163-0.0093  1.2536  1.2339 -0.3251  0.5353 -0.3197 -0.8711 -1.2363 -0.6511  1.6257 0.0913  0.6147 -0.5108  1.4818  0.7554 -0.3573  0.2901  1.9580  1.1486 -1.0507 0.8533 -0.4667  0.4017 -0.2958 -1.3575  1.6106 -0.1584  0.3790 -1.3659  0.3711-1.4533  1.6274  0.4612 -0.1936  0.6430  0.7840  0.6470 -0.1030 -0.0883  0.9078 0.5480  0.5556 -0.9327 -1.5060  0.3498 -0.2189 -0.8157  0.4640  0.9918  0.1536-0.6035 -1.1391 -1.9268 -1.2533 -0.3209  0.5090 -0.0997 -1.1278 -0.8528 -0.6946... [the output was truncated (use n=-1 to disable)]
#>