Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.1896  0.1423 -0.6657 -0.3848 -0.3446 -0.8688  0.7387 -0.2121  0.4614 -0.8468
#>  0.9694  0.1028 -0.0233  1.5701 -0.5586 -0.2163  1.5558 -0.3068 -0.6669  1.0840
#>  0.8609  0.0316  0.2327  0.3380 -0.7392  0.7867  0.8315 -0.4818 -0.6131  0.5489
#>  0.1523  1.6700  0.0521 -0.4959 -1.1781  2.2915  1.0935  1.9675  0.3699  0.6703
#>  0.3845 -0.5147 -0.4331  0.9277  0.9220  0.3356  0.6562  3.2472  0.1044  0.8790
#>  0.1059 -1.1738 -0.0153  1.3150 -1.2149  1.5235  1.2542  0.1234 -0.5936  0.0824
#>  1.6064 -0.1161 -0.2630 -1.0804  1.7294  1.0712  0.3999  0.3253  0.3592  0.3327
#> -0.2274  1.6316  0.6002  0.3067 -0.6796  0.8614  0.3898 -0.0313 -0.8836 -1.4237
#>  0.2477  0.8678 -1.0249 -0.7100  1.2030 -0.4195 -0.5587 -0.1424  1.6674 -0.4266
#>  1.3462 -0.8278 -0.4759 -1.1250 -0.4499 -0.4954 -1.4353 -0.2113 -1.3631  0.9288
#>  0.5752 -0.7482  1.6690 -1.2782  2.1417  0.5994 -1.6672 -1.0408  0.0237  2.7737
#>  1.9800 -0.1654 -0.7788  0.5773 -0.2913  2.9616  0.4077  0.0108 -0.0878  0.1759
#> -0.1017  0.9577  0.5423  1.8005 -0.3740  1.2178  0.2218  0.8073 -0.4134  0.9072
#> -0.2479 -0.0400 -0.2981  0.5309  0.8770 -0.9530  1.8266  1.2817 -0.2827 -0.0345
#> -1.0616 -1.3954 -0.0900 -0.8192 -0.5208  0.2754 -0.7975  0.0712 -0.1671  0.1917
#> -1.3425 -0.3843  1.3532  1.6029 -0.5144  0.9014 -1.3362  1.3960 -1.0101  2.8473
#>  2.1382 -0.9575  0.0611  1.2730 -1.1858 -0.5111  0.0545  0.2023 -0.5893  1.3608
#>  0.8358 -0.3642  0.2439 -1.2212 -0.3787  0.3034  0.6432 -0.4959  0.8923 -1.7276
#> -1.3613  1.0855 -0.0218 -0.7821  1.7147  0.1558  0.4430 -0.2470 -0.8179 -1.1722
#>  1.4510 -0.1315  0.0284 -1.6782  0.7110 -0.8087 -0.4680 -0.5330  0.2456  0.5620
#>  1.6448  0.3812 -0.9793 -1.0265 -0.3831 -0.5430  1.3858  1.5569 -0.1564  1.1492
#> -1.6779 -0.2288  1.4636 -0.3625 -0.0297  1.5356 -0.2485  1.7250  0.0340 -0.4998
#> -0.6541  0.4075  0.3093  1.4747 -0.5549  1.6495  0.6866 -0.9872  0.4368 -1.9997
#>  1.3158  1.0273  0.2474 -0.2991 -0.4011 -3.0767  1.3491 -2.2113  0.4144 -0.1927
#>  2.2935 -0.2902 -0.1710 -0.9830 -1.3515  0.6451 -0.0473 -0.2418  0.1960 -0.1740
#> -0.5066  0.4889 -0.7468  0.2557  0.1849 -0.4020  0.4954 -0.3215  0.3372 -0.9882
#> -0.6807  1.0357  0.7313 -0.7285 -0.7024 -0.3107 -0.4055 -0.2363  0.0633  0.8403
#>  1.1555 -0.5990  0.7219 -1.4270  0.8919  0.4160  0.4443  1.5912 -0.1820 -0.0819
#> -0.7589  1.4631  1.0285 -2.6050 -0.4987 -1.6802 -1.1450 -0.0659 -0.4208 -1.5072
#> -1.2096  0.0371 -0.5123  0.9621 -0.4292  2.3245  0.3343 -0.4453 -0.6405  0.5855
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]