Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.0073 -0.2470  0.3935  0.8951  3.5847  0.7390  0.2025 -0.3993  0.6610 -1.0197
#>  0.0966 -0.7471  1.4031 -0.5893  0.3067  0.1450  1.3427 -0.5554  0.3419 -1.2341
#>  0.2466  0.2377  1.4483  1.3428 -0.1626 -0.1915 -0.8707 -1.6185  0.4098  0.2364
#>  2.1328  1.7239  0.7756 -0.4856 -0.7450 -0.6215 -0.1312 -0.6940  0.2565  2.3401
#> -0.9860  0.3390 -0.2407 -0.8218  0.0119 -0.0795  1.6154  1.9040 -1.1239 -0.0491
#> -0.7041  0.9390  0.6200  0.6053 -0.3897 -1.5667 -0.1396 -1.0318 -0.3710  1.9475
#> -0.2791 -0.0542 -0.2385 -0.2470  0.9430  0.0084 -1.1371 -0.2894  0.0860  2.0937
#> -1.9326  1.4231  1.2527 -1.0377 -0.2751  0.7045  0.2655  0.9116 -1.6634  0.7085
#>  0.8698  0.4033  1.1153 -1.5301 -2.0786 -0.4054  0.3117  0.0019 -1.0720 -0.9359
#> -2.0683 -1.2287 -2.2947 -0.3370 -0.1736 -0.7281 -0.1305 -0.6936  1.2512  0.1577
#> -0.0103  0.1889 -1.4098 -0.8876  0.7914 -0.9162 -0.2777  1.7068 -0.4336  0.9884
#>  1.3646  0.5318 -1.0621  0.7583 -1.3035 -1.4554  0.0388 -0.7140 -0.1920  1.3428
#> -0.5575 -2.0095  0.4120  0.4911  0.9398 -1.1370  0.5789 -1.5030 -0.3722  1.4175
#>  1.6369 -0.5253 -0.7980 -0.2562 -0.1085  0.6596 -0.2839  0.2299 -2.0786 -1.3685
#>  0.9887 -0.4340 -0.2982 -0.9823 -0.9521 -0.8572  1.0698  0.2369  0.9708 -1.1605
#>  1.1277 -0.6447 -2.2490 -0.9129  1.3609 -1.1850  0.3525  0.6168 -0.3791  1.5584
#>  0.9951 -0.1879 -0.9588 -0.0665 -1.6838 -0.3951  0.0540  1.1821 -2.5940 -0.8599
#>  1.0655  0.3302  0.6998  1.5301 -0.7093 -0.3518 -0.9381  0.4383  0.8040  0.7226
#>  1.2312  0.0496  0.7910 -0.0332  2.1623  0.3008 -1.0661  1.7234 -0.2040 -0.8167
#>  1.3582 -2.2070  0.8932  0.9996 -1.9588  0.0622  1.2665  0.0210 -1.2858 -0.1799
#> -2.9498  0.0180  1.0051 -0.4081 -1.9847 -0.1982  0.1762  0.3912  0.5140 -0.1139
#>  0.6542 -0.4677  0.2891  0.6002  1.0336 -0.4990  0.2213 -1.8331 -1.1975 -0.6640
#> -1.0664 -1.1301 -1.2373 -0.5508  0.9607  0.0664 -0.3465  0.0174  0.6513  0.0661
#>  0.1590 -0.2899 -1.1918 -0.0299 -0.2954 -0.9461  3.4682 -2.0181  0.1464  2.0149
#> -0.1906 -0.9905  0.8467 -0.3878  0.7446  1.0145  1.3510 -0.5463 -3.0474 -0.5609
#>  2.4299 -0.1049  0.6860 -0.9751  0.1939  0.8254  1.3813  0.5871  0.1112  0.1640
#> -0.8211 -0.2929  0.7251 -1.2577  0.5135  1.3137 -1.1763  0.8526  0.5281  1.0860
#>  0.9460 -1.5822  0.3313 -1.0051  2.0028  1.0318 -1.4410  2.3345 -1.4520 -2.1052
#>  0.4507 -0.7531 -0.5057  0.0954 -0.5908 -0.8229 -0.6490 -0.3948 -1.3616  0.6793
#> -0.6662 -0.1676 -1.8954 -0.1566  2.6392  1.1797 -0.6634 -0.7859  0.3942 -1.4952
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]