Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5386 -1.0307 -1.2615 -0.4335  0.3020 -0.4219 -0.1211 -1.0067 -0.9536  0.8931
#> -0.1346  0.6061 -1.9100  1.1195  0.9410 -0.7977  0.3256  0.4594 -0.4880  0.7815
#>  1.0629  0.8733 -1.2590  1.3207 -0.0100  2.2110 -1.8044  0.0264 -0.3012  1.3560
#>  0.2704 -0.0173  1.3994  1.2372 -0.2001  0.4334 -0.8036 -0.0073  0.3038 -0.2518
#>  0.1453 -1.7228 -0.8896 -0.3467 -1.7421  1.5601 -1.6626 -0.4087  2.0143  0.5068
#>  1.1537 -0.5866 -0.6666 -0.7557 -0.7809  1.0631  0.5913  2.7526 -1.2795  0.9775
#>  0.8748 -0.8696  1.9020  1.1367 -0.5137 -0.7300 -1.6259  0.5550 -0.9495  0.7094
#>  1.2829 -0.3472  1.3881 -0.4432 -1.2597  0.1460 -0.5071 -0.5391  0.3705 -1.0290
#> -0.7239  1.1269  0.0640  0.5826  0.7943  0.0690  0.1969  0.0306 -0.1088 -0.5618
#> -0.2237 -0.1756 -1.2910 -0.2612 -1.6694 -0.4233  0.1383 -0.4875  0.7908 -0.2240
#>  1.8198  0.5731  1.0371 -1.8721  0.3093 -1.2748 -0.8164  1.3966 -0.8880 -0.2755
#>  1.0533  0.8713 -1.5987 -1.5986 -0.1446  0.5827  0.5507  2.7468 -0.6940  0.7778
#> -1.3508  1.5601 -0.1775 -0.0066 -1.6470  0.1083  0.1806 -1.3991 -1.2243 -1.5552
#> -1.4892 -0.4004  0.8139 -1.2190 -1.0379 -2.6568 -1.2176 -0.3395 -0.3025  1.3392
#>  0.3220 -1.0715 -0.2531  0.0335  0.3097 -0.9728  1.6837 -1.3450 -0.3965 -1.0846
#>  0.1916  0.5235  0.6333  1.3363  1.3001  1.0959 -0.4993 -0.3825 -0.2823  0.4946
#> -1.9193 -0.2148  0.0402  0.2798  0.6025  0.6839  0.1183 -0.2679  1.2578 -0.7159
#> -1.4639  1.3120  0.5926 -1.0783 -3.0700 -1.1928  0.8376 -0.7395 -1.2339 -1.0249
#>  0.2834  1.2973  0.8377  0.3049  0.2392 -0.4703 -0.1222 -0.9949 -0.4656 -0.4945
#>  0.0161  0.5134 -0.2275  0.6433 -0.4865  2.4436 -0.6141  1.5112  0.9209 -0.5798
#>  0.1891  0.8364  1.2367 -1.3520  1.0157  0.4031  0.2317 -0.3650 -0.0272  0.1154
#>  1.1281  1.7854  1.6604 -0.6118 -0.5090 -0.6817  2.4928 -0.0452 -0.3039  0.4493
#> -2.4239 -0.2975 -0.0965  1.1585 -0.0810 -0.5701  0.1046 -0.3358  0.5295  0.3494
#>  1.6928 -0.6767 -0.1100  1.6352 -1.5756 -0.4276 -0.0401 -0.5164  0.9382 -0.0925
#> -0.2261  1.1585  0.0044 -0.1739 -0.8068  1.1025 -0.5096 -0.6739  1.1954  0.0563
#>  1.8120  0.0546 -1.2133 -0.2552 -0.1174 -1.2079 -0.7313 -0.4307 -0.1399 -1.0286
#>  0.4102  0.3857  0.5635  0.9829  0.8176  1.4976 -1.7981 -0.6327 -1.3159  2.2531
#> -0.6767  0.9127 -0.6144 -0.1018  0.3990 -1.6461  2.0782 -0.4000 -0.7841  2.1450
#> -1.3449  1.9914 -0.9792 -1.5902  0.4803  0.2039  0.1683 -1.3401  0.0760  1.3145
#> -0.0996 -1.1388  0.7321 -1.0030 -0.1741 -2.2039 -1.6337  0.5093 -1.6345 -0.3132
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]