Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.2830 -0.4873  0.5085  0.3850  0.6723 -0.7060  2.3817 -1.1614 -0.5275  2.0085
#>  1.2667  0.4227 -1.3822 -0.7476 -0.1762 -1.3975 -0.9304 -0.1415 -1.7157 -1.4153
#> -1.3052 -0.4027 -0.7798  1.8776  0.7139 -1.4986  0.6782  0.2230 -1.2551  0.4011
#> -1.2634  0.7362 -0.4947  0.7170  0.5996  0.1617  0.7291  1.7551  1.4738  0.0898
#>  0.1962  0.5856  1.5068 -0.0899  2.2529 -1.0695 -0.4592  0.3260 -1.0602  0.7051
#> -0.2560 -0.0907  1.1854  0.4755 -0.3598 -2.2520  0.9044 -0.1709 -1.0646 -0.2845
#>  1.1366  0.2032  0.8516 -0.1093  0.6582 -0.9427  0.2670 -1.3719 -2.1837 -0.3219
#> -1.1379  0.7023  0.3062 -0.3597  0.3792  0.1993  1.9483  0.3639 -1.7183  0.7538
#>  0.2111 -0.8799  1.5661 -1.2769 -0.2521  0.8547  1.4843 -2.4578  0.8653 -0.3767
#>  0.4588 -0.2351 -0.3436  0.1126  0.8149 -0.7157 -0.5442  1.8504  0.1129 -0.3002
#>  0.0280  0.0661  1.0415 -1.8621  1.2550 -2.2430  1.5041  0.1291 -0.1949  1.7753
#> -1.4225  0.0498 -1.8347 -0.2468 -1.1179 -0.1579  1.0117  1.1008  0.1293  0.1703
#>  1.0024 -0.4845  0.1192 -0.3126  1.7822  0.3481 -0.7764  0.2495  0.2706  0.7251
#> -1.5091  0.5246 -0.4008  0.8710 -0.7933  0.7201 -0.0001 -1.0088 -0.3805 -1.5550
#> -0.7466 -1.9043  0.2300  2.0016 -1.9244 -2.2192  1.4554  0.0050 -0.1100  0.2427
#> -0.7252 -1.1626  1.0426 -0.3251  0.1707  0.5396  1.3166 -1.9897 -1.3763 -0.8959
#>  0.0542 -0.5962 -0.9810  0.5157  0.8164  0.9609  0.4140  0.1518 -0.2004 -0.7686
#>  0.4269  1.3737 -0.6778 -0.9904 -0.6879 -0.8327 -0.7574 -0.3025  0.1085 -1.3325
#> -0.2675  1.5189  0.8500  0.6693 -1.4864 -0.6050 -0.4418  0.1185  0.4614 -0.1868
#>  1.0849  0.4011 -0.9632  1.7188 -0.8338  0.2358  0.1372 -0.0811  0.4490  2.7482
#> -1.0630  0.8449 -1.0040 -0.6649  0.5998 -0.9089  0.2028 -0.4198  0.0660 -0.2703
#>  1.0802 -0.9735 -0.0200 -1.4126  1.8983  0.6004 -0.0597 -0.0762 -0.3082 -1.7273
#> -1.3452 -0.3706  1.7043  1.0260 -1.4453 -0.4402  1.5259 -0.6532  0.3388  1.2633
#> -0.2445  1.8675 -1.4696  0.3023 -1.6782  1.4962  0.9038  0.2479  0.3139 -1.0893
#> -2.7692  1.1330 -0.3976 -1.5069  0.1969  0.5842 -0.6362  0.0040  0.1068  0.5367
#>  1.9436  1.3692  1.5383 -1.1703  0.5792  0.7368 -0.0712  0.4559  1.1857  0.9746
#>  0.0790 -1.3244  0.2457  2.0050 -0.2677  0.4553 -1.9313  0.1907 -0.4195 -0.6692
#>  1.2319  0.1883 -0.5482 -0.3267  2.9302  2.3143  0.5232  1.6036  1.0604  0.4202
#> -0.6914 -1.3410  0.7265  0.5287  1.5564 -0.5351 -0.9788  0.2395 -0.5410  0.7291
#> -1.1083  1.3115 -1.4045 -1.5762  0.2168 -0.2358  0.4785  1.8240 -0.5395 -1.2047
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]