Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.0243 -1.3886 -0.7197  0.3864  0.9747  1.7485  0.5869  0.8015 -0.5725  0.2822
#> -0.5038  0.0660 -1.9018  0.5524  0.5166 -0.1456 -0.0920  0.1839  0.6089  0.4606
#> -0.8836  0.8136 -0.0420 -0.8126 -0.3800  0.0697  0.5343 -0.5651 -0.0618 -0.7980
#>  0.5106 -1.0779  1.2285 -0.5112  0.0280 -0.9169 -0.3770  2.4139 -0.4506  1.6020
#> -3.9573  2.1106  0.6488 -0.2942  1.9589  0.6747 -0.4591  0.9795  0.2548 -0.7040
#>  2.3276 -0.9151 -1.5298 -0.0335  1.9982 -1.8013  0.6993 -1.4125  1.0678 -0.5441
#>  0.3094 -0.0634  0.8126  2.3983  1.9091 -1.6453  0.3656  0.6590  1.3563 -0.8200
#>  0.0912 -0.6317  0.2048  0.3944  0.5989  1.1606  0.4960  1.7781  1.1993 -0.5648
#> -1.3209  0.4783 -0.3425 -0.4001 -0.0982 -1.1266 -1.0556  1.2805 -0.3971  0.1660
#>  1.4807  0.3204 -0.0888 -1.3811 -0.6371 -0.2537 -1.5559 -1.3382  1.0520  0.9165
#> -1.4811 -1.6676 -0.4687 -0.8289  1.0546 -1.3234 -0.5865 -0.4664 -0.9237  1.2923
#> -0.2387  0.2224 -0.1203  0.8515 -0.4303  1.1753  0.3367  0.5898  0.3165  0.4326
#> -0.6364  1.2416 -0.6548 -0.5300 -0.3607 -0.3184  1.8592  1.2066 -1.5569 -1.7900
#> -0.6298 -0.4292 -0.3413 -1.4684  1.4088  0.5927  0.5187 -0.4920 -0.5475 -2.7966
#>  0.2486  0.2263 -0.0806  0.1379 -0.9781  0.1216 -0.8686 -0.8092  2.5457  0.7017
#> -0.2289 -0.9653  0.2597 -1.3661  0.1813 -0.1348  0.1253  0.2832  0.7101 -0.1959
#> -0.4839  0.7928  0.5102 -1.6952  0.4114 -0.9888  0.2519 -0.0752  0.3717 -0.7058
#> -0.4628  1.1405  0.5652  0.5450 -0.5139 -0.4371 -0.0960 -1.6452  1.6288  0.4854
#> -0.1059  0.5084  0.6433  0.5997  1.5350 -0.8783  0.6447 -0.3508 -0.2428  0.6326
#> -0.3831 -0.6679  0.0242 -1.1600 -0.6287  0.4104  1.0862  1.6226  0.4354 -0.1938
#>  0.0910 -0.6679  0.1938  0.3846 -1.1381  0.0369  0.2453  0.3954  0.2730 -1.6508
#>  0.7007  1.3639  0.0738  0.4354  0.5534  0.3142 -0.3664 -0.5786  1.2865  0.9770
#> -0.5917  0.6842  0.2852 -0.6755 -0.4501 -1.0945 -0.6620  1.2631  0.3575 -0.2142
#>  1.1682 -0.0756  0.7807 -1.4833 -1.3126  0.4131  0.1624  0.3743 -0.0278  0.0411
#>  0.1275  0.2760 -0.0379  1.1077  0.5034 -0.6229  0.4837  1.1094  0.3929 -0.1099
#> -0.4228  2.2286 -1.6475  1.0716 -1.2181  0.1369  0.2464  0.1016 -0.9079 -0.7724
#>  2.0110 -0.7438 -0.7675  0.0760  1.7213  0.8297 -2.1225 -0.0415 -0.7123 -0.1972
#> -1.6767 -0.1755  1.7895  1.2881  0.4331 -0.4192  0.7572  1.5774  0.3353 -0.9085
#>  1.1065 -1.2900 -1.3546  0.4085  0.1279 -1.7906  0.4281 -0.3481 -1.4288  0.7137
#>  1.2524 -0.2611  0.2964  0.1869  1.1096  0.1567 -0.6128 -0.5941 -0.6102 -0.2509
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]