Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.6367 -0.8256  1.0173  0.4963 -0.1358  0.3316 -1.1917 -1.4656 -0.0849 -0.9699
#>  0.5337  0.1680 -0.3068 -1.2912 -0.6326  0.7114 -0.1828  0.9161 -2.0421 -1.9189
#>  0.4643  0.7411 -1.7726 -0.3290 -0.1870 -1.3260 -0.5294 -0.8193  0.0250  0.6417
#> -0.6018 -0.0287 -1.4458  0.9021 -0.9531  0.0538  1.2981  0.4136  0.4173  1.4147
#>  0.9023  0.2954 -1.2290  1.3417  0.7716  0.0155 -0.4026 -0.0899 -0.8024  2.3220
#>  1.0322 -0.3204  0.4297 -0.9972 -0.1329 -0.0522  1.9034  0.6641 -0.2588  0.2226
#>  1.2743 -0.9863 -1.3960 -0.9271  0.1491  1.0279 -0.0136  0.0249 -0.5191 -1.2069
#> -0.3863  1.3361 -0.4276 -0.3298  0.1152 -1.5805 -0.5238 -0.5339  1.9375 -1.2904
#> -0.1207 -1.0428  0.8923 -1.3193 -1.1486  1.2769  1.3823  1.2450  1.6633  0.6624
#>  0.2800 -0.3915  0.1971  0.5192  0.0207 -0.3134  0.7032  2.4264 -2.7792  1.8488
#> -0.0916  0.1106  0.4693  0.0690  0.2172 -0.7792 -0.0350 -0.7472  1.4384 -1.6677
#>  0.8410 -0.3180  2.3857 -0.9325  0.9467  1.3093  0.3319 -1.7845 -0.1270 -0.2145
#>  1.0098  0.5211  0.2833 -0.2214 -0.9217 -0.2560 -0.2248 -0.2645 -0.5927 -0.2379
#>  1.3181 -0.8984  0.8291  0.9253  0.9599 -1.3781 -1.2243  1.7686 -0.1544  1.2521
#>  0.6789 -0.3227  0.5277  1.7740  1.7829  1.6399 -2.4166  1.2546  0.8701 -0.1225
#> -0.9212  0.4398 -0.1715  0.2791 -1.9287 -1.1875 -1.1755  0.8505 -0.9198 -0.1238
#> -2.1762 -0.7040 -1.3640 -1.1083  0.6672  0.5589  1.4835 -1.1290  1.5212  0.6136
#>  0.4786  0.6457 -0.5270  0.1554  0.7130 -0.4496  2.0330 -0.6951  0.3572  0.2507
#> -0.0395  0.5846 -0.7949  0.5414  0.9363 -0.2466  1.4081  0.5493  0.2044 -1.4661
#>  1.6383  1.2887 -0.7069  0.9333  0.1673  1.7474  0.5684 -2.2659  1.0150 -0.5884
#>  1.0313  0.2290  1.1900 -1.0924  1.9231  0.0584 -0.7819  0.4116 -0.9683  0.5781
#>  0.2707 -0.5556  1.7838  0.3584 -1.7144 -0.1190 -0.1401  0.5074  1.8272 -0.5108
#> -1.0628  2.0625 -1.3349  0.5256 -0.4919  0.9708 -1.7341  0.9789 -0.3173 -0.3259
#>  0.6189  0.9191 -0.0217 -0.8085 -0.2774  2.0489 -0.0489 -0.0931 -1.8453  0.5577
#> -1.2615  0.9138 -0.8355 -0.1952 -0.1533  0.0135 -0.2877 -0.6386 -1.9984 -0.4245
#>  0.7072  0.7738 -1.7004 -0.6148  0.1939 -1.8354  1.3249 -0.1382  0.0522  1.4541
#> -0.8817  1.8765  1.1167  1.2157  0.6548 -0.3264  1.7991  0.3475  0.7077 -1.0980
#> -1.9031  0.5427  0.7395 -0.5530 -0.4650  0.7123  0.1179  0.1805  2.0030  0.6657
#> -3.1274  1.2942 -2.3046 -0.5384  0.6885  0.2109  0.6306  0.6219 -0.7307  0.6795
#> -0.4488 -0.5523 -0.7859 -0.9278 -1.1016 -0.1612  0.4723 -0.5055 -1.5903  0.1340
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]