Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-1.1834 -0.8307 -0.9972  0.5007  0.4768 -0.8345  0.8198 -0.8301  2.6006 -1.2823
#>  0.0666 -0.7159 -1.8196 -1.5294 -0.9880 -1.4160  0.4070  0.5359  2.1046 -0.4677
#>  0.0903 -1.0553  0.4944  0.1829 -0.2641 -0.4079 -1.3497  1.7222 -0.5087  1.5156
#> -2.2194  0.4061 -0.0581  1.2390 -0.0981  0.0074  2.1333  1.5163  0.2465  0.2291
#> -0.8275  0.8364  0.0291 -0.2733 -1.4954 -0.3957 -0.2984 -0.4345  1.0223 -1.1092
#> -0.4767  0.2483  0.6487  0.1082  0.2673 -0.6590 -0.6806 -0.5644 -3.0498  0.3049
#> -0.8711 -0.7949  0.4634 -0.2605 -1.2261 -0.9679 -0.4277 -0.2766  0.3014  0.9988
#> -2.3251 -1.5977  0.3140 -0.8579 -1.2555 -0.4110  0.5122  0.3057  0.2180  0.2558
#> -0.8691  0.9297  0.2826  1.5235  2.1746  1.1274  1.9997 -0.2118 -0.0188  0.7249
#>  1.3232 -0.6880  0.6085  0.7496  0.4717 -1.0421 -1.1427  0.0648  0.5356  0.8564
#> -1.6989  1.1377 -2.7065  1.7080 -0.8946 -1.1624  1.4811 -0.8350 -0.1068  0.8488
#>  1.1905 -0.6297  1.2549  1.4561  1.0354 -1.2034 -2.0957 -0.6425  1.8968  1.1373
#> -0.0302  0.4286 -0.6191 -0.3321  1.2937 -0.6212 -0.1675  1.4340 -0.4241  2.7561
#> -0.5559 -0.6766 -0.7740 -1.4766 -0.4829  0.7766 -2.1202 -0.5490 -0.4970 -0.2443
#> -1.8387  1.0524  0.7940 -0.5152  1.1661 -0.1564  1.0941  0.7884  0.2874  0.4763
#>  1.0395 -0.4455  0.4491  1.0113  0.1252 -0.3141 -0.0082  1.1480  0.3301  0.8884
#>  0.3232  0.1767 -0.1061  0.2466  0.4113 -0.6303  0.1330 -0.1473 -0.2608 -0.8104
#>  0.2492 -0.7989 -0.2981  0.7371  0.6663  1.9673 -0.1405  0.7679  0.9249 -2.0467
#> -1.1041  0.6213 -0.2693 -0.3520 -0.3473  0.1054 -1.7054 -0.6219 -0.5212  0.5953
#>  1.3040 -0.3785  0.7068  0.0242 -0.1164  0.5910 -0.3874  1.3949  0.3714 -0.9868
#>  0.5220  0.0511 -0.2188  0.2864  0.1461 -0.7788  0.5360 -0.7055  0.8804 -0.8159
#> -0.8519  0.3875 -0.3524  1.5649  1.1323  0.6864 -1.7215 -0.0405 -1.9537  0.1145
#>  0.1145  1.5309  0.5366  0.9104  0.6306 -0.4446  0.6876 -0.1611 -0.3437 -1.2315
#>  0.3062 -1.3065  0.9870 -1.3889 -2.7052  1.3369  0.9826 -0.7980 -0.1986 -1.4328
#>  1.4372 -1.6052  0.5923 -0.2428  0.9664 -1.2978  0.7571  0.4400  0.4389 -0.6299
#>  1.2600  1.0423 -1.8036  0.0518  0.6824 -0.5560 -0.6450  0.6876 -0.2473 -0.3548
#> -0.8164 -0.4363 -0.1989  0.3473  0.4315  0.2247  0.3983 -1.2318 -0.1293  1.5298
#> -0.9492 -1.6003 -0.8633 -0.2492 -1.0395  1.1222  0.9562  0.3648 -0.2372  1.4446
#>  1.2577  0.1338 -0.3257  0.7397  0.3046 -0.3489  0.6650  1.0799  0.1701 -0.3482
#> -0.1270 -1.2342 -1.0755  0.1128  1.4712  0.2525 -1.3594  1.9416 -0.2873  1.0294
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]