Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 6-4.8963e-01 -4.6032e-02  1.1099e+00  1.2991e-01  1.1150e+00 -8.4112e-01
#>  1.3278e+00  3.9578e-01 -4.7505e-01 -2.0945e+00  1.6771e+00  2.5081e-01
#> -6.1959e-01 -6.9444e-01 -1.0943e+00  8.4966e-01  1.1877e+00 -7.1341e-01
#>  5.9460e-01  1.0859e+00  5.1299e-01  3.3193e-01 -3.0133e-01 -1.9106e+00
#> -1.8127e+00  1.6373e+00 -1.9544e+00 -1.4107e-01 -2.0113e+00  2.7575e-01
#>  1.6169e+00 -1.1563e+00 -4.5134e-01 -9.5703e-02  1.0506e+00 -4.9454e-01
#> -1.5623e+00 -7.5619e-01 -6.0841e-01  2.8195e-01  6.1505e-01  6.4247e-01
#> -1.6863e-01 -7.4566e-01  1.8029e-02  1.7657e+00  3.3351e-01  3.9945e-01
#>  1.2158e-01  2.3556e+00 -1.1760e-01 -3.9971e-01  3.6550e-01  3.6606e-01
#>  1.2771e+00  1.1525e+00 -2.1599e+00 -1.9412e-01  6.4840e-01  6.9386e-02
#> -2.9263e-01 -1.0276e+00 -7.9427e-01 -5.9293e-02 -8.4165e-01  4.7883e-01
#>  9.0980e-02 -2.5968e-01  3.9365e-01  1.6773e+00  5.3472e-01 -8.0174e-01
#> -1.7620e+00 -1.8711e-01 -9.4356e-01  6.9917e-01 -8.8245e-01 -1.3410e+00
#> -2.7895e-01 -1.6054e+00 -1.1660e+00  4.6062e-01 -1.5366e+00 -6.7081e-03
#>  7.1504e-02  1.2829e+00 -6.4900e-01  3.9112e-01 -4.1275e-01  2.0883e-02
#> -7.9383e-01  5.9285e-01  1.7250e+00  1.0003e+00 -9.6004e-01  8.9258e-01
#> -1.6029e+00 -6.6425e-02  2.1268e+00 -4.1136e-01  1.2353e+00 -1.1236e+00
#> -3.4746e-02  9.5520e-01  3.7267e-01  1.2701e-01  1.5624e-01  1.3125e+00
#>  7.7337e-01  6.5590e-01  1.5214e-01 -1.5334e+00 -3.6632e-01 -2.7117e-01
#>  1.3285e-01  1.4268e+00  2.3214e+00  2.0448e+00 -1.0891e+00  1.7531e-01
#> -1.0669e+00  3.0627e-01  2.2406e+00 -1.5254e+00 -1.1095e-01  3.7948e-01
#>  1.4610e-01  1.8090e+00  5.1519e-01 -2.1852e-01  8.0514e-01  1.8098e-01
#> -7.3834e-01  7.6131e-01 -2.9061e-01 -1.8877e+00  2.0529e+00  1.1242e+00
#> -1.0904e-01 -4.4613e-01 -1.1310e+00  9.3151e-01  8.5533e-01 -1.3163e+00
#>  1.1987e-01  9.8304e-01  1.2568e+00  1.0209e+00  2.5487e-01 -7.5287e-01
#> -1.0237e+00  5.5857e-01  3.4515e-01  5.8053e-01  1.2455e+00 -2.9166e-01
#> -7.8607e-01  3.4704e-02 -8.5123e-01 -1.1410e+00 -9.5984e-01 -4.1555e-02
#> -2.8976e+00  1.9974e-03  5.2411e-01  9.5189e-02 -9.8860e-01 -4.7244e-01
#>  6.8375e-01 -5.6605e-01 -1.0786e+00  8.5261e-01 -5.2646e-01 -2.5432e-02
#> -1.8316e+00  1.2379e+00 -3.1440e-01 -4.1087e-01  8.9752e-01  1.8153e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]