Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.5732 -0.3110  0.7293 -0.5468  0.8314  2.1647 -0.0362 -1.1807  0.5785  0.3868
#> -1.1411 -1.0270 -0.0113 -1.0385  0.2011 -0.7171  1.6702 -0.0403 -0.4555 -1.1282
#> -1.1338 -0.4795  0.1309 -1.7253 -0.0962 -2.1419  0.7131 -0.0070 -2.6906  0.8460
#>  0.7309  0.1417  0.4706  1.1489 -0.4139  0.0965 -0.2927  0.5580 -0.1521  0.4033
#> -1.2384 -0.7301  0.4607  0.1754  0.0628 -1.5095 -0.5459  2.2341  0.2746  1.5567
#>  0.3345 -0.2247  0.2387 -0.9027  1.2656  0.3804 -1.1095 -0.0632 -0.2398 -0.0279
#> -0.7435  1.9179 -0.4272  0.2052 -0.1299  0.3426  0.5200 -2.6083 -0.3467 -1.0300
#> -2.1239 -0.8722 -0.6727  0.7608  0.6741  0.6710 -3.1947 -1.4586 -0.1653 -1.2217
#> -0.6769 -0.5960 -0.1928  0.6230 -0.3608  1.1152  1.3624 -0.6803 -0.1050  1.0042
#> -0.1001 -0.5827 -0.0799  0.7049  0.5478  1.0381 -0.4502  1.0485  0.0375 -0.0135
#>  0.9526 -0.2509 -0.2946 -0.5693 -0.0878 -0.4221 -1.3923 -1.5588 -0.6964 -1.1354
#>  1.6444  0.0698 -1.4403  2.3463  0.0566 -0.6933  0.7791 -3.2416 -0.1658  0.4884
#>  0.2854  1.5350 -0.1970 -0.8266 -1.3388  1.1145  1.2707  0.4805 -1.3855  0.0219
#> -0.4777 -0.1661 -1.5242 -0.7172 -0.6026 -0.4491 -1.5737 -2.1681 -0.8394  1.8660
#> -0.8093  1.8655 -1.3884 -0.0883  0.6091 -0.5132  1.1409 -1.0626 -0.3957  0.5630
#>  0.1716 -0.7183 -0.4147 -0.3561 -0.0758 -1.0497 -0.3009 -0.7979 -0.6101 -1.8065
#>  0.1190  1.4202  1.0685 -0.8871  1.1306 -0.6938 -0.5440 -1.6249  0.6574  0.5808
#> -0.6185  0.0544 -2.8086  1.1698 -0.1091 -0.5448  0.9981  0.6927  2.3155 -1.6879
#> -1.2487 -0.8063 -0.2489  0.2811  0.1151 -0.5131 -1.8780 -0.5573  0.5505  0.7255
#> -0.3913  0.0514  0.4565  1.3496  1.3544 -0.7302 -1.8260 -0.7240 -0.5498  1.1038
#> -0.3859 -1.2561  1.4780 -0.1667  0.8226 -0.0308  1.8582 -0.2012 -0.0870  0.7713
#> -0.0656  0.8216 -1.6940 -1.0484 -0.3110 -0.2733  1.0824  0.5463  0.4070 -0.6766
#>  1.6234 -0.2154  1.1282 -0.0128 -0.6562 -0.5163 -1.6046 -0.6618 -0.6529  1.0296
#> -1.3405  0.7137  1.0605 -0.3395  0.3055  0.5946  1.8739  0.1773  1.8577  0.1009
#> -0.8563 -0.4252  1.5768  0.0611  0.7135  0.3884 -1.4025  0.5007 -0.2813  1.0076
#> -0.0941 -1.7674 -0.0466  1.5993  0.7307 -0.3864 -0.1687 -0.8520  1.0947  1.3583
#>  0.2060  1.5227 -0.6719  1.7604 -0.4997 -0.9993 -0.3696  1.4368  1.8417 -0.3211
#> -0.9329  0.6065 -0.4007  0.8170 -0.4768  0.6136 -0.7694 -0.2755 -0.2012 -1.4127
#>  0.5978 -0.2771 -0.2350 -1.5329 -0.7855  1.5693  1.3123  0.5008 -1.6835  0.6596
#> -0.9907 -0.2757 -1.5746 -0.0118  0.1004  1.9579 -0.8001 -0.6992  0.1741 -1.0683
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]