Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.4239 -0.5060 -0.0685  0.8549  2.3621 -1.2925  0.0311  0.1275  1.7983  0.6535
#> -1.9021 -0.9197  0.0087 -0.1652  0.3546 -0.1004 -1.0624  0.6528 -1.0211 -0.1740
#>  0.7410  0.1509  0.0408 -0.9556 -1.1828  0.5820  0.5591  0.5559  0.0214 -1.0381
#>  0.7983  1.4925 -0.7279  0.5412 -0.6095  2.2405  0.1317 -0.1719  0.4419  0.7171
#> -0.8411  0.3465 -1.6742  0.5673  1.6819 -0.8655 -0.2761 -0.1179 -0.3039  0.4329
#>  0.6599  1.0778 -0.1547 -0.3588  0.7130 -0.7832  0.0486  1.5571  1.4597  0.4227
#>  0.4194  0.2884  1.5229  1.0555 -0.0112 -1.1946  0.6937  1.5771  0.2529 -0.5586
#>  1.3114 -0.8105 -0.6845 -0.7336  0.1311  0.8733  1.7924  0.0158 -0.8906  0.9398
#> -0.6085 -1.4775  0.6976  0.7665  1.1619 -0.4755 -0.5701  0.8052  0.8485  0.3216
#> -1.1867  0.6922 -0.0291  1.9655  1.9324 -1.0231 -0.2475  0.8086  0.3284 -1.1765
#>  0.7424 -0.7175 -0.3331 -0.6144  1.2293  2.1546  0.6782  1.7380 -0.6232 -0.6360
#>  0.4137  1.5775 -0.5228 -0.7818  0.3125 -0.2267  0.5056  0.8874  1.3401 -0.7305
#> -0.7577  0.8803 -0.0568  1.0064 -0.5384 -0.1681  0.4229  0.0097 -0.4072  1.5097
#> -2.0812 -0.2552 -0.7149  0.9723  0.4575  0.2062 -0.4020 -0.8135 -1.0310 -0.4863
#> -1.5834 -0.8905  0.8791 -0.7605 -1.1725 -1.8507 -1.0908  0.5064 -0.4473  1.0980
#> -1.7450  1.8597  0.0587  1.8571 -1.1970  0.3424 -1.4217 -1.1943  0.0336 -0.9086
#>  0.3364  0.4770 -1.2373 -1.3862  0.1850 -0.7548  0.8911 -0.9988  1.2016 -0.7714
#>  1.7306  1.3966  0.9892 -0.2892 -0.9862  1.4523 -0.7258 -0.5434  1.8328  0.2910
#> -2.1066  0.6623  0.5846 -0.2091 -0.1164  0.4183  2.4118 -0.4557 -0.9186 -2.2659
#> -0.4298 -1.8547 -1.2803  0.4078  0.3126 -2.1307 -0.4352  0.1705  0.3447  0.4468
#> -0.3829 -0.1503 -1.7807 -0.6722  1.1296  0.9062  0.7036 -1.5080  0.1025 -0.8652
#>  0.0938  0.2346 -0.6362  0.4377 -1.0486  0.8825 -0.4916 -0.3670 -1.2143  0.8475
#>  0.0806  0.5126 -0.6631 -1.4201  0.4413 -1.2116 -0.4685 -1.3537 -1.6633 -1.1041
#>  0.4054  0.9296  2.4399 -1.4248 -1.4413 -0.1736 -0.3009  0.2101 -1.1454  0.3249
#> -0.7263  0.0529 -0.9699 -1.6964  1.4177  0.4924  0.5250 -1.1306 -1.6086 -0.1592
#> -1.0983 -2.7479 -0.3220  1.0979  0.6634 -0.6141 -1.5607  0.5036 -0.9316  1.7806
#>  0.6975  0.3536  0.0097 -0.3757  0.5127 -0.4199 -0.5605  0.2315 -0.4184  0.8231
#> -1.6081 -1.1752  1.6388 -0.7227 -0.2234 -1.8669  0.3815  1.0956 -0.0314 -0.3447
#> -0.6449  0.9173 -0.7022 -0.2918  0.2150  0.7576 -0.1219 -0.8401  2.8083 -0.3494
#> -0.4198 -0.9344 -0.3559  0.8667 -0.3286 -0.8934 -0.3765  0.7742  1.2524  1.8911
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]