Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10-0.1959 -0.9557 -0.9333 -0.7497 -1.2595 -0.3841  0.0597 -0.3891  1.2656  0.1732
#> -1.8826 -0.1636  0.7025  0.7813 -1.0804 -1.4226 -0.8488  0.7551 -0.1180  0.4580
#> -0.5447  0.3785  0.0211  1.6554 -1.0166 -0.9188 -0.0705 -0.3180 -0.3014 -1.1981
#>  0.9708  0.2371 -1.5848  0.8809 -2.3951  0.3742 -1.6655  0.2285 -0.9261  1.6304
#> -0.7441  0.9543  0.6815  0.4257  0.8335 -0.0550 -1.2198 -0.2481  1.5449 -1.0358
#> -1.8707 -1.4920 -0.0412  0.3737 -0.7544  1.1349 -0.0347  0.7469 -0.0623 -0.4047
#>  0.2216 -1.7803 -0.4100  0.6104 -0.2137 -0.9547 -0.6752  1.4997 -0.9876  0.0565
#> -1.4674 -0.7113 -0.2286  1.0833 -0.0672  1.6671  0.5930 -0.4177  1.1451 -0.6182
#> -0.4809 -1.0277 -0.1236 -1.5840 -0.1486  1.1866  1.0617  0.8845  0.0943 -0.7045
#> -2.9395 -0.3331 -1.4852  0.3970 -0.5944  0.5123  0.5137  0.6318 -0.6854 -1.0804
#> -0.0395 -0.6315  1.1090 -0.8152 -1.0312 -0.5153 -0.5646 -0.6111 -1.0488 -1.2002
#>  1.4705 -0.6718  1.5765 -0.0024 -0.4671 -1.5720  0.9473  0.7509 -1.1833  0.0983
#> -1.4971  0.6050  1.7357  0.6581  0.3012 -0.1013  1.8944 -1.9345  0.3022  1.8667
#>  0.2652  1.5465 -1.0566 -0.3845 -0.4172 -1.1119 -0.9223 -0.0177  0.4181 -0.2542
#> -1.2055 -0.6424 -0.6767  0.7400 -0.2961  0.5564  1.0281  0.7771  2.1210  0.7747
#> -0.0171  1.9214 -0.1572 -0.9770 -0.7424  0.6680  1.1684 -1.0410 -0.9915  1.3275
#> -0.1325 -0.1433 -0.8269 -0.0572  1.2497 -0.4987 -2.4392 -0.0580  0.4711 -0.5651
#> -0.2685  0.6711 -1.3169 -0.5633  0.6805 -1.0155 -0.6383 -0.6152 -0.4855 -1.1734
#>  1.1939 -0.4419 -0.4084 -0.3624 -1.0613  0.5091  0.8732 -1.3795  1.4907 -0.7721
#>  0.1145  0.8111  0.5903 -0.2994  1.0953  0.7421 -0.7114 -0.5324 -0.1989 -1.0725
#>  1.1997  0.7954 -0.4245 -1.0610  0.3413  0.7662 -0.2073 -2.1859  0.6167 -0.1157
#> -1.1566 -0.3981 -1.2527 -1.0708  0.4111 -1.7679 -0.1448  0.1308  0.6108  0.3056
#>  0.6118  0.3633  1.3159  1.5631 -0.6947 -0.5417  1.0816 -0.5872 -1.1572 -1.0427
#>  0.1958  0.6038  0.7227  0.0485  0.2570 -0.0915  0.5151 -0.2737 -1.1701  1.4684
#> -0.4527 -2.1033  0.7346 -0.2803 -0.9147 -0.4465 -0.6754 -0.5886  0.5454 -0.1940
#>  0.2237 -0.4176  0.3416  1.6399 -1.6909 -0.0744 -1.3408 -0.3997  1.0276  1.9272
#>  1.2664  0.2879 -0.6262  0.9227  1.3465  1.2566  1.2657  2.0240  0.8316  0.4751
#>  0.3020 -0.1311 -0.3022 -1.8832  1.0369 -2.2832  1.5164  0.3965 -0.3838 -1.5202
#>  0.5559 -0.2955 -0.9590 -0.1785  0.7058 -0.7922  0.0115  0.8471  0.1516 -0.2617
#> -1.2080 -0.3451  1.5972 -0.5100 -0.6857  0.9268  1.0033 -0.1947 -1.3012  1.4373
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]