Flattens a contiguous range of dims into a tensor.

For use with nn_sequential.

Usage

nn_flatten(start_dim = 2, end_dim = -1)

Arguments

start_dim

first dim to flatten (default = 2).

end_dim

last dim to flatten (default = -1).

Shape

Input: (*, S_start,..., S_i, ..., S_end, *), where S_i is the size at dimension i and * means any number of dimensions including none.

Output: (*, S_start*...*S_i*...S_end, *).

See also

Examples

if (torch_is_installed()) {
input <- torch_randn(32, 1, 5, 5)
m <- nn_flatten()
m(input)
}
#> torch_tensor
#> Columns 1 to 10 0.7560 -2.2702 -0.3309  1.5430  0.3181  0.6472 -0.8968 -0.7577  1.2092 -0.7812
#> -2.9833  3.1933  1.0188  1.2087 -0.5740  0.6842  0.0103 -0.6510 -2.3857 -1.2641
#>  0.2743  0.6831  1.0162  0.6561  0.7328  0.2560 -2.1972  1.2709 -0.5826  0.3802
#>  1.0352  0.7943  0.1942  2.4063  0.7112 -1.0282 -0.7984 -0.6103 -2.0002  0.8131
#>  0.9153  0.4423 -0.8704  1.4634 -0.9504 -0.0401  0.0423  1.8535  1.7030  0.6453
#>  1.3175 -0.0103 -2.3672 -1.8348  0.2914  0.4603 -0.5191 -1.1672 -0.1679 -0.7710
#> -0.6445 -0.2680 -0.4697 -1.7980 -1.4941  2.5633  1.4720  0.8057  0.0666  0.0946
#>  0.8519  1.1563  2.7594 -1.5383 -0.4483  0.6128 -0.1267 -0.6962 -1.3308  1.0505
#>  1.6477 -0.6239 -0.9407  0.0415  0.2087  0.0113  0.8241  0.8914  0.0170 -1.6049
#> -0.3499  2.3432  0.1668  2.1904 -0.2436 -1.4888 -1.4951  0.3688 -0.2227  0.4850
#> -0.1021  1.0697  1.2794 -1.0922  0.4779  0.4316  1.3622  0.9541 -1.0037 -0.2763
#> -0.6246  0.1202 -0.7652 -0.0639  1.2339 -0.7199 -0.7972 -0.9931  0.4580  2.0635
#> -0.2732 -1.4262 -0.1815  0.5873  1.6224  2.7023 -1.1495 -0.7254  0.0137 -1.6834
#>  0.2641 -1.0748  1.4947 -0.4340  0.9507  1.4452  0.4817 -0.2445  0.9866 -0.9237
#>  0.6604 -0.5477  0.4396 -0.4329  1.7488 -0.1252  0.3675  0.9005  1.6571 -0.0278
#> -0.0261  0.6476 -0.5090 -0.2455  0.6426 -0.5655 -1.0654 -0.1893  0.7139  0.9771
#> -0.2493 -1.3428  0.2749 -0.2537 -1.3353  0.9013 -0.0240  0.7629 -0.0610 -1.8086
#> -0.2547  1.7450 -1.8809  1.2345  1.0470 -0.5050  0.4051 -0.9022  0.6760  1.6150
#>  1.1086  0.5550  0.0418 -0.0090  0.4965  0.0531  1.2174  1.6017 -0.4074 -1.1282
#>  0.2801 -1.3712 -0.2074 -0.9782 -1.5624  2.0371  0.7663  1.7544 -0.1520 -0.1488
#>  0.0316  0.9093  0.0029  1.0738 -1.0493  0.9351  0.5692  0.6474  0.9282  1.0728
#> -0.3428 -0.6433 -0.7907 -0.6171  0.4029  0.8790  1.1121 -1.5268 -0.2809  0.8142
#>  1.3083 -0.2187 -2.1541  0.5084 -0.3623  0.0123 -0.9872 -1.5555  0.7507  1.3483
#> -0.4577 -0.3726  0.8692 -0.8177  0.7447  0.6227 -0.2650 -1.1304 -1.2267  1.0705
#> -0.7437 -0.0567  0.9838 -1.3419  0.6634 -0.2496 -0.4427 -1.6795 -0.3003  1.7777
#>  0.7045 -0.2349 -0.4019  1.1735  0.3034 -0.2561  0.3320 -0.9833 -0.6106  2.0013
#>  0.6878 -0.1234 -0.8210  0.4898 -0.5578 -1.2518  0.6952  0.2886 -0.0640  0.3889
#>  0.5003 -0.4531 -1.1189 -1.6787 -0.1595 -0.8194 -1.1372 -0.6231  1.2330  0.3033
#> -1.3515  0.7787  1.4412 -0.1686 -0.7951  0.4805  0.4227  0.5610 -0.3515  1.0690
#> -0.1339  1.3257  0.0133 -0.9972 -0.9098 -2.0038  1.0846  0.3920 -0.3241 -1.5521
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{32,25} ]