RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0794  0.0789  0.1247 -0.4615  0.9647 -0.0644  0.4401 -0.4098 -0.8003
#>   0.5180  0.2285  0.1498 -0.8606 -0.2680  0.1038 -0.0618 -0.1125 -0.7881
#>  -0.3697 -0.0190 -0.7993 -0.4271  0.8824 -0.4947  0.1302  0.0567 -0.0433
#> 
#> Columns 10 to 18  0.3478  0.3985  0.6812  0.5608  0.2449 -0.2245 -0.1833  0.3448  0.1878
#>   0.4061 -0.3900 -0.5003  0.0883 -0.0721  0.0380 -0.4709 -0.7572  0.1675
#>   0.8414  0.7868  0.9524 -0.0100 -0.0409 -0.8107  0.6164 -0.4292 -0.8881
#> 
#> Columns 19 to 20 -0.4645 -0.4991
#>   0.0758  0.2655
#>  -0.8195 -0.8869
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2606 -0.4569 -0.2104 -0.4862 -0.2230 -0.1431  0.6373  0.2952  0.0015
#>   0.4263 -0.3802  0.1485 -0.0640 -0.1871 -0.3991  0.2478  0.2643  0.2838
#>   0.5477 -0.1879  0.3430  0.0622  0.4972 -0.3451  0.2969 -0.3678 -0.1920
#> 
#> Columns 10 to 18  0.5891  0.2211  0.4588  0.0030  0.0481 -0.0233 -0.3206 -0.1073  0.3237
#>   0.5465  0.3769  0.1982  0.3703 -0.3404  0.1892 -0.3143  0.0634  0.0567
#>   0.5131 -0.0172  0.2320 -0.2452 -0.2773 -0.0297 -0.7587 -0.5911 -0.0088
#> 
#> Columns 19 to 20  0.5609  0.3288
#>   0.1338  0.2664
#>  -0.4529 -0.1847
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3590 -0.4618  0.0351 -0.1778  0.0849 -0.1875  0.3200 -0.4084 -0.5008
#>   0.2160 -0.1758 -0.2566 -0.3338 -0.4270  0.2426  0.2798 -0.1914 -0.0401
#>   0.3049 -0.0817  0.1855 -0.3322 -0.2725 -0.3455  0.1719  0.4600 -0.0108
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2804 -0.2681 -0.1663 -0.1852 -0.1313  0.3309  0.0749  0.0093  0.1435
#>   0.7804  0.3458  0.1838  0.3661 -0.5759  0.0383  0.4529 -0.2443 -0.2188
#>   0.4201  0.1652 -0.2895  0.6936 -0.7797  0.7330  0.3262 -0.1541 -0.2073
#> 
#> Columns 10 to 18  0.1262 -0.4712 -0.4419 -0.2728 -0.5923 -0.6169 -0.4415 -0.0313  0.2827
#>  -0.1065 -0.1553  0.4864  0.0827 -0.5035  0.0606  0.0275  0.5786  0.0191
#>   0.2362  0.0836 -0.2764 -0.1975 -0.5896 -0.3695 -0.1727 -0.2020  0.5161
#> 
#> Columns 19 to 20  0.7836 -0.4527
#>   0.4998 -0.2859
#>   0.4391  0.5321
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5398 -0.3859  0.2564 -0.0376 -0.3766 -0.2255 -0.1021 -0.1159 -0.6281
#>  -0.0764 -0.2084  0.1754 -0.2924 -0.3440 -0.2807  0.2932 -0.1780 -0.2500
#>   0.4978 -0.0175 -0.2085 -0.1663 -0.2580 -0.2290 -0.3163  0.0295 -0.4537
#> 
#> Columns 10 to 18  0.6832 -0.0293  0.3094  0.1850  0.1539 -0.0085 -0.1113 -0.2713  0.0319
#>   0.6300  0.0215  0.4085 -0.0551  0.4542  0.1821 -0.0957 -0.3319  0.2019
#>   0.3576  0.0140  0.1192  0.1110 -0.0644  0.0602  0.0671  0.4748 -0.1480
#> 
#> Columns 19 to 20 -0.0484 -0.3258
#>   0.0682  0.1967
#>  -0.0858 -0.4025
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>