RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0967  0.1489 -0.0959  0.5797  0.3785 -0.2021 -0.7980  0.3053  0.1843
#>  -0.1143 -0.6840 -0.8942 -0.6228 -0.3665 -0.5787  0.8640  0.7376 -0.4592
#>  -0.6201 -0.2936  0.0082  0.6765  0.7561  0.5933 -0.5355  0.3461  0.8331
#> 
#> Columns 10 to 18 -0.7426 -0.3871  0.2543 -0.2732 -0.1844 -0.8923  0.0899  0.1456 -0.2423
#>  -0.0595 -0.3189  0.7126 -0.3229  0.8529  0.0190  0.8334  0.5752 -0.6043
#>   0.6227  0.3776 -0.7403  0.0354 -0.0708  0.3707  0.4445 -0.0215  0.6161
#> 
#> Columns 19 to 20  0.7818  0.5044
#>   0.1670 -0.8759
#>   0.5037 -0.1844
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1892  0.5616 -0.1640  0.2648  0.5850 -0.2579 -0.3202 -0.4865 -0.0095
#>  -0.3644  0.4721 -0.1311  0.7653  0.4378  0.3996 -0.6275  0.0329  0.5965
#>  -0.0817 -0.2652 -0.2706 -0.0762  0.3069 -0.0454 -0.1647 -0.5307  0.1979
#> 
#> Columns 10 to 18  0.0507 -0.2082 -0.1331  0.2911  0.2465 -0.4303 -0.0790 -0.2926 -0.1002
#>  -0.1689  0.1575 -0.0331  0.2997 -0.0440  0.0379 -0.7864 -0.1360  0.0149
#>  -0.2516  0.2386  0.4662  0.0429  0.4360 -0.2954 -0.1112 -0.1386 -0.0092
#> 
#> Columns 19 to 20  0.0752 -0.4384
#>   0.2191 -0.0554
#>  -0.1911 -0.6736
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3023 -0.2808 -0.1948  0.1790  0.1540 -0.4187 -0.1109  0.4069  0.6312
#>   0.3547 -0.2781 -0.0828 -0.3018  0.5085 -0.6500 -0.3520 -0.4971 -0.0070
#>   0.0845  0.2891 -0.5550  0.1253  0.5191  0.0397 -0.5007  0.0938  0.4580
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0899 -0.2668  0.8330  0.5370 -0.6343 -0.2770 -0.4611 -0.4930 -0.0211
#>  -0.1489 -0.3450  0.6120  0.3678  0.7029  0.0916  0.5174 -0.4117 -0.7150
#>   0.2016  0.0084 -0.0678 -0.1749  0.2979 -0.6042  0.4544  0.5378 -0.1114
#> 
#> Columns 10 to 18  0.2462 -0.2703  0.6181 -0.4772  0.2232 -0.0782  0.2253  0.3155 -0.0823
#>   0.8420 -0.2251 -0.1442 -0.1411 -0.0989 -0.1383  0.2341 -0.5215  0.5583
#>  -0.1835  0.4185 -0.2710 -0.0628  0.3882 -0.3590 -0.4242  0.0301  0.5127
#> 
#> Columns 19 to 20 -0.2712 -0.0604
#>  -0.7274 -0.3536
#>  -0.4918 -0.3270
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0517  0.2745 -0.3450 -0.0609  0.4183 -0.1824 -0.3268 -0.1065  0.4317
#>  -0.0918 -0.0234 -0.3601 -0.0486  0.5176 -0.4196 -0.1359 -0.6597  0.0518
#>  -0.0694 -0.0249 -0.2779  0.4993  0.0919  0.0240  0.2950 -0.5826  0.0694
#> 
#> Columns 10 to 18 -0.5416 -0.0364  0.5415  0.2190 -0.0055 -0.2884 -0.1616  0.0426 -0.5690
#>  -0.5421  0.2865  0.1599  0.2554  0.3624 -0.3182 -0.2364 -0.0594 -0.2020
#>  -0.0173  0.3086  0.0734 -0.0190 -0.0704 -0.4375 -0.0989  0.1285 -0.0494
#> 
#> Columns 19 to 20 -0.0399 -0.1025
#>   0.1281 -0.7231
#>   0.0885 -0.2358
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>