RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4641 -0.7712 -0.1833  0.1225  0.2104  0.0367 -0.0510  0.3262 -0.0619
#>  -0.5222  0.0420  0.2229 -0.3491  0.1934  0.6127  0.7164  0.8187  0.1706
#>   0.2361 -0.7161 -0.3399  0.3065  0.8878 -0.7936 -0.2267  0.2741  0.5383
#> 
#> Columns 10 to 18  0.0225  0.1380  0.3392 -0.7653 -0.2351 -0.6603 -0.2827 -0.8221  0.5145
#>   0.3747 -0.8621  0.8915  0.7007  0.0120  0.0144 -0.7481 -0.0206  0.3809
#>   0.6115  0.2255  0.8835  0.2980 -0.6831 -0.4030  0.8362 -0.9345 -0.5333
#> 
#> Columns 19 to 20 -0.5304 -0.1240
#>   0.4056  0.7686
#>  -0.0684 -0.3008
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3555  0.2600 -0.2652 -0.0963  0.4225  0.1128 -0.0858  0.1876 -0.1076
#>  -0.2286  0.6953  0.1362 -0.3512 -0.0946  0.2231  0.1182  0.4302  0.0882
#>   0.2619  0.5298  0.4228 -0.2022  0.1458 -0.0255 -0.4108  0.3289  0.2842
#> 
#> Columns 10 to 18 -0.4457  0.3512  0.2985 -0.5666 -0.1910  0.4764 -0.3385 -0.2742 -0.0655
#>  -0.0390 -0.3179  0.2871  0.5041 -0.4392  0.1055  0.1103 -0.1552 -0.1501
#>  -0.2683  0.3257  0.5379  0.0184 -0.2019  0.2151  0.3888 -0.3075 -0.0915
#> 
#> Columns 19 to 20 -0.0369  0.4629
#>  -0.3713  0.2500
#>  -0.0294 -0.1344
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0221  0.1815  0.2028  0.0010  0.2565  0.2604  0.0555 -0.2588  0.0138
#>   0.1376 -0.2395  0.2249  0.2994  0.0231  0.1243  0.4184 -0.3753  0.2618
#>  -0.0464  0.0087  0.4831 -0.0405  0.0653 -0.1121 -0.0689 -0.0658  0.1782
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2857  0.7499 -0.7432 -0.4088 -0.0559 -0.3466  0.5556  0.7065  0.2409
#>   0.6703  0.4297 -0.0461 -0.6096 -0.0202 -0.0275  0.5187  0.3147  0.0130
#>   0.5649 -0.2677  0.2230  0.2340 -0.2944  0.1309  0.3369  0.7670 -0.2497
#> 
#> Columns 10 to 18 -0.3867 -0.3806 -0.0062 -0.7220  0.1345 -0.1560  0.2672  0.0650 -0.0744
#>  -0.1833 -0.3342  0.4345 -0.8862  0.3322  0.3802 -0.0067  0.6365  0.6707
#>  -0.3026 -0.1789  0.3430 -0.1422  0.5009 -0.4513 -0.0174  0.1220 -0.1466
#> 
#> Columns 19 to 20 -0.1811 -0.0757
#>   0.3351 -0.8487
#>  -0.1592 -0.2465
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1438  0.3855  0.1340 -0.4608 -0.0562  0.1262  0.2681 -0.2479  0.2880
#>  -0.3146  0.5297  0.3622 -0.3267  0.0661  0.0224  0.0884  0.3301  0.2227
#>  -0.1468  0.3125 -0.0231 -0.4317  0.1121 -0.0432 -0.0819  0.0912  0.2574
#> 
#> Columns 10 to 18  0.1511  0.1372  0.6067 -0.5516 -0.3920  0.0673  0.0292 -0.2414  0.2444
#>   0.1157  0.0977  0.3878 -0.2692 -0.2647 -0.2836  0.1354 -0.3745 -0.3476
#>   0.0298  0.3636  0.2403 -0.2088 -0.3010  0.1817  0.4147  0.0824 -0.0985
#> 
#> Columns 19 to 20  0.1973 -0.0680
#>   0.0769 -0.0564
#>   0.3542 -0.4791
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>