RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6379  0.1701 -0.5194  0.4892  0.3338 -0.8842  0.0669 -0.0994  0.7735
#>   0.0087  0.8038 -0.1664 -0.5587  0.1503  0.2393 -0.6139  0.4705 -0.2544
#>  -0.7197 -0.8243 -0.5876 -0.4554  0.7278 -0.6614 -0.3453  0.0182  0.2097
#> 
#> Columns 10 to 18 -0.4666  0.3140  0.2902 -0.4316  0.5666  0.4352 -0.7450  0.1522  0.2693
#>   0.4249 -0.1173  0.6198 -0.4079  0.4232 -0.4182  0.6386  0.3998 -0.1977
#>   0.4717 -0.0705 -0.9812 -0.2998 -0.0750 -0.1248 -0.3925 -0.3143  0.1046
#> 
#> Columns 19 to 20 -0.8723  0.2716
#>   0.1297  0.3436
#>   0.2947 -0.0534
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0456 -0.2184  0.0299  0.3598  0.5817 -0.5528 -0.1228 -0.2041 -0.1089
#>   0.3579  0.4381  0.1788  0.5408 -0.4901 -0.7670  0.0317  0.0777 -0.4169
#>   0.0285 -0.0753  0.2706  0.6980 -0.0177  0.1195  0.0376  0.2281 -0.2917
#> 
#> Columns 10 to 18  0.2775 -0.4619 -0.6673 -0.3185  0.5757 -0.2291 -0.5796 -0.3228  0.3357
#>   0.5484  0.5521  0.4571 -0.2820  0.1023  0.3057  0.1629 -0.6055 -0.2904
#>   0.4282  0.3472 -0.6766 -0.0453  0.2270  0.2526 -0.3846 -0.4583  0.7108
#> 
#> Columns 19 to 20 -0.5195 -0.3882
#>   0.0841  0.0839
#>  -0.1006 -0.3557
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.6263 -0.3172 -0.1626  0.1093 -0.2145 -0.3666  0.2423 -0.5080  0.3586
#>  -0.4042 -0.2408 -0.0430  0.5144  0.6329 -0.6042 -0.0048  0.3816  0.1532
#>   0.5428 -0.3980 -0.3842 -0.0867 -0.1972 -0.1192  0.4984  0.0101  0.2851
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6935 -0.1051 -0.1059 -0.5822 -0.1887 -0.1275 -0.5008 -0.4580  0.2349
#>   0.0788 -0.3262 -0.0733 -0.7665 -0.7955  0.0268 -0.7030  0.4179  0.2278
#>   0.4043  0.2032 -0.0720 -0.5229  0.0626  0.0028 -0.6429  0.2446 -0.0476
#> 
#> Columns 10 to 18 -0.1713 -0.2931 -0.3526  0.4746  0.6676  0.3189  0.0723  0.1936  0.4416
#>  -0.5430 -0.5525 -0.0930 -0.1187  0.3213 -0.0503  0.0984  0.8444 -0.3177
#>  -0.6532 -0.5172  0.4437 -0.0120  0.2369  0.5376 -0.1365 -0.0433 -0.4401
#> 
#> Columns 19 to 20 -0.6470 -0.0302
#>  -0.4404  0.4412
#>  -0.7127 -0.0387
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3071  0.0245  0.0217  0.3952 -0.0717  0.3455  0.2631 -0.0160 -0.5803
#>  -0.1341 -0.3060  0.2049  0.4955  0.3884 -0.1562  0.0365 -0.1799  0.2450
#>  -0.1286 -0.1824 -0.0489  0.4726  0.2853 -0.2321 -0.3116 -0.1634 -0.3471
#> 
#> Columns 10 to 18  0.2394  0.2733 -0.1205 -0.0386 -0.0131  0.2992  0.1042 -0.4349  0.1066
#>  -0.2009 -0.0897 -0.6742 -0.3660  0.4796 -0.0993  0.0102 -0.2629  0.4095
#>   0.3339  0.2251 -0.6246 -0.2354  0.0686  0.4484 -0.2822 -0.4437  0.1667
#> 
#> Columns 19 to 20 -0.2441  0.0481
#>  -0.4608 -0.3050
#>  -0.1698 -0.2514
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>