RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0099  0.8873 -0.1116  0.5582  0.2236 -0.8063 -0.5214 -0.7094  0.0030
#>  -0.6134  0.6402  0.2909  0.1591  0.3425  0.1394 -0.1715 -0.4264 -0.1925
#>  -0.2075  0.8212 -0.2350  0.6117  0.8437 -0.8979  0.1299  0.4098 -0.1091
#> 
#> Columns 10 to 18 -0.1063 -0.5816 -0.6736 -0.8061 -0.3460  0.0017 -0.7643 -0.6852 -0.0715
#>   0.8909  0.5398  0.2206  0.0341  0.4578  0.4309 -0.3563  0.7488  0.2405
#>  -0.1761 -0.2920 -0.9140 -0.4591 -0.4857  0.7049 -0.2167 -0.0460 -0.6598
#> 
#> Columns 19 to 20 -0.4293  0.2680
#>  -0.3719  0.4223
#>   0.0940  0.9169
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3324 -0.4207  0.0305  0.3984  0.0503 -0.5823  0.1092  0.5120  0.2031
#>  -0.3039  0.3983  0.2596 -0.0226  0.1558 -0.2363  0.2077 -0.0720  0.0413
#>  -0.6505 -0.1879  0.2960  0.1649  0.0949 -0.6909  0.2089 -0.2326  0.4585
#> 
#> Columns 10 to 18  0.0796 -0.3511  0.4411 -0.2902 -0.2510 -0.5497  0.1241 -0.1329 -0.3761
#>   0.3596 -0.0047 -0.1558 -0.3505  0.0080  0.1363 -0.0167  0.1671 -0.3121
#>   0.1444 -0.1915 -0.3381 -0.2206 -0.4520  0.0327 -0.3149 -0.0472 -0.2376
#> 
#> Columns 19 to 20 -0.7047 -0.0313
#>  -0.3081  0.0800
#>  -0.2463 -0.2978
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4079  0.2917 -0.2918  0.3921  0.0511 -0.2014  0.2033  0.0481 -0.1901
#>  -0.4303 -0.0670 -0.4264  0.0645  0.3152  0.0342  0.0529  0.3867 -0.1832
#>  -0.4909  0.2774 -0.2381  0.3101  0.4226 -0.1731  0.1337  0.1107 -0.5018
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8149 -0.3475  0.4000 -0.1804  0.5313  0.2956  0.3946 -0.0228 -0.6228
#>   0.0342  0.5787  0.5131 -0.3002 -0.0482  0.4752 -0.1411 -0.1442 -0.5239
#>  -0.5904  0.1925  0.6303 -0.2594 -0.4145  0.0472 -0.4394 -0.2307  0.1172
#> 
#> Columns 10 to 18 -0.4365 -0.5798  0.2951  0.1726 -0.2926  0.2333 -0.0807  0.4533  0.2642
#>  -0.6537  0.4116 -0.0362 -0.0832  0.1841  0.1278  0.0342  0.4623 -0.4961
#>   0.1162 -0.4211  0.4740 -0.2485  0.1174 -0.4846  0.4786  0.1228  0.1390
#> 
#> Columns 19 to 20 -0.6162 -0.3449
#>  -0.5088 -0.3846
#>   0.0963 -0.2808
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6454  0.2634 -0.5312  0.4242  0.4415 -0.2678  0.0404  0.2957  0.6067
#>  -0.7700  0.4239 -0.3339 -0.0677  0.2237 -0.0442  0.3492 -0.0533 -0.1367
#>  -0.7737 -0.1304 -0.1595  0.1469  0.0405 -0.3013  0.3678  0.0060 -0.0411
#> 
#> Columns 10 to 18  0.1754 -0.0503 -0.2273 -0.1933 -0.5074  0.0287  0.1542 -0.1830 -0.5688
#>   0.5005  0.2507  0.1192 -0.3769  0.0007  0.6051  0.5483 -0.1710 -0.4696
#>   0.4363  0.0470  0.2786 -0.4432  0.0068  0.0723  0.3499 -0.1739 -0.2832
#> 
#> Columns 19 to 20 -0.0214 -0.0146
#>  -0.0557  0.3255
#>  -0.3888 -0.2729
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>