RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2249 -0.5577 -0.3009 -0.4637 -0.6397 -0.4488 -0.6692 -0.8016 -0.0698
#>   0.4453 -0.9609  0.2623 -0.3803  0.5264 -0.4049 -0.8967 -0.0039 -0.3156
#>  -0.5458 -0.7069  0.8648  0.2045  0.7063 -0.7025 -0.2655  0.0702  0.3081
#> 
#> Columns 10 to 18 -0.6916  0.8883  0.1480  0.2981  0.6952 -0.5384 -0.8319 -0.1959 -0.8632
#>  -0.9362  0.2779  0.5784  0.7914  0.4215 -0.1272 -0.3004 -0.5692  0.3363
#>   0.4600 -0.5724  0.8666 -0.2371  0.5806  0.3176 -0.5486  0.9668  0.6385
#> 
#> Columns 19 to 20 -0.0021 -0.5065
#>  -0.3767  0.6971
#>   0.2871  0.7296
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5339  0.1036  0.2759 -0.4129 -0.1800  0.2679 -0.0560  0.1490 -0.5948
#>  -0.5665  0.2095  0.0283 -0.3985 -0.4345 -0.5291  0.4511 -0.1668 -0.0370
#>   0.3108 -0.4567 -0.0868 -0.2810 -0.1743 -0.6445 -0.0537 -0.0774 -0.4668
#> 
#> Columns 10 to 18  0.0758 -0.4195 -0.0614  0.1044  0.0513 -0.5154 -0.2501 -0.0699  0.3685
#>   0.1288  0.1753  0.5442  0.1649  0.1585 -0.6613 -0.3584  0.2814 -0.2041
#>   0.0516  0.1591  0.4101 -0.0238 -0.0063 -0.7487 -0.4581  0.0492 -0.4343
#> 
#> Columns 19 to 20 -0.2198  0.0936
#>  -0.2072 -0.3608
#>  -0.3517 -0.3486
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1073 -0.2270  0.4937 -0.3504 -0.2845 -0.2052  0.0666 -0.1323 -0.1369
#>   0.3829 -0.1378  0.4751 -0.2849 -0.6704 -0.0468 -0.3835  0.2119 -0.5066
#>  -0.1643  0.3763  0.1636  0.0562 -0.4611  0.0309  0.0496  0.1880 -0.2777
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4382  0.3527  0.8788 -0.0011 -0.1215 -0.7132 -0.4003  0.2167  0.1302
#>   0.6872 -0.6399  0.3774  0.7477 -0.4645  0.6521  0.0758 -0.3665  0.7359
#>  -0.6905 -0.3278  0.1429 -0.3505 -0.2616 -0.4839 -0.3683  0.6959 -0.5444
#> 
#> Columns 10 to 18  0.0848  0.1216  0.1184  0.0235 -0.5787  0.2194  0.3643  0.2878 -0.0846
#>  -0.0220 -0.0523  0.1921 -0.6307 -0.7977  0.7906 -0.3145 -0.0817 -0.6312
#>  -0.2598  0.3416 -0.3765  0.1604 -0.4370 -0.2098 -0.4568  0.3521  0.4189
#> 
#> Columns 19 to 20  0.5244 -0.2329
#>   0.9143 -0.3009
#>   0.3553 -0.4460
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1282  0.4115 -0.2042 -0.4658 -0.3705 -0.0589 -0.1598 -0.2124 -0.3172
#>  -0.5100 -0.1198 -0.1025 -0.1653 -0.0866 -0.2309  0.2014  0.0126 -0.1463
#>  -0.2020  0.0316  0.1237 -0.1705  0.0389 -0.1105 -0.1841  0.2265 -0.0591
#> 
#> Columns 10 to 18 -0.2517 -0.5239 -0.0610  0.1796  0.1608 -0.2343 -0.0372  0.3311 -0.4673
#>   0.2684  0.4834  0.2034  0.1677  0.1929 -0.3546 -0.2541  0.2636 -0.1285
#>   0.1184 -0.5301  0.1306  0.1305  0.0908 -0.4976 -0.1386  0.3243 -0.1578
#> 
#> Columns 19 to 20 -0.0812 -0.4301
#>   0.0504 -0.3604
#>  -0.2570 -0.1548
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>