RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1207  0.2611 -0.5084 -0.3467 -0.8021  0.2311  0.2564 -0.6574 -0.2451
#>   0.8315 -0.4214 -0.3159 -0.5470  0.7895 -0.6093 -0.4343 -0.1334  0.4206
#>   0.3350 -0.6341 -0.2711 -0.0237  0.8395  0.0510 -0.6476  0.5944 -0.5971
#> 
#> Columns 10 to 18  0.2170  0.0220  0.3842 -0.5600 -0.4562 -0.5591 -0.5762 -0.1722 -0.9209
#>   0.1922  0.1895 -0.5314 -0.8723  0.7297 -0.8596 -0.4922  0.7427  0.2153
#>   0.8988 -0.3478 -0.2031  0.0532  0.5350 -0.2715 -0.1444 -0.3268  0.8681
#> 
#> Columns 19 to 20 -0.5318 -0.8989
#>   0.9764 -0.4440
#>  -0.1543  0.6733
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4049 -0.4608 -0.2807  0.4198  0.3313  0.6510  0.7849  0.2340  0.3933
#>  -0.6356  0.1409 -0.4906  0.0928  0.3070  0.7646  0.4516 -0.3472  0.3558
#>   0.0902 -0.4056 -0.7900  0.2810  0.5057  0.2046  0.1942 -0.1337  0.0033
#> 
#> Columns 10 to 18  0.6559 -0.7392 -0.0612  0.6534 -0.5611 -0.1714  0.3548  0.3939 -0.0945
#>   0.4724  0.3450  0.2959 -0.2063 -0.2247  0.1187 -0.0252  0.2905  0.1287
#>   0.3512 -0.1618 -0.3062 -0.0021  0.1430 -0.1565  0.1796  0.5522  0.0375
#> 
#> Columns 19 to 20 -0.4958  0.7416
#>   0.2793  0.0733
#>  -0.2078  0.2324
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1877 -0.1816 -0.5898 -0.4439  0.2548 -0.2786  0.5639 -0.3519 -0.3301
#>   0.3687 -0.3241 -0.3660 -0.2842  0.0367  0.4814  0.0948 -0.4875 -0.0371
#>  -0.0069 -0.1542 -0.4865 -0.0923  0.1952  0.4144  0.4028 -0.2488  0.1977
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1438  0.2873  0.4280 -0.7491 -0.0060  0.4595  0.3263  0.4011 -0.6952
#>  -0.4313  0.0955  0.1725 -0.4069 -0.5577  0.0052  0.1516  0.5795 -0.6583
#>  -0.4344 -0.5251  0.1010  0.3383  0.5300 -0.6911  0.5475  0.0696  0.4202
#> 
#> Columns 10 to 18 -0.1542  0.3629  0.3895 -0.2519  0.5190 -0.6827 -0.0403 -0.5716  0.9348
#>  -0.1497  0.5548  0.0384  0.4710  0.4384 -0.5356  0.3715  0.0885  0.7734
#>   0.3381  0.1023 -0.3359  0.5232 -0.1521  0.5500 -0.8372  0.4367 -0.5602
#> 
#> Columns 19 to 20  0.6709  0.1124
#>   0.6790 -0.4410
#>  -0.3319 -0.1476
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1875 -0.0258 -0.4184  0.1941 -0.0381  0.1675  0.4663 -0.1295 -0.1429
#>   0.0021 -0.0443 -0.2934  0.2278 -0.0150  0.4249  0.4645 -0.0311 -0.2499
#>   0.1803 -0.5234 -0.1506  0.1965  0.3645  0.4084  0.4105  0.0987 -0.1185
#> 
#> Columns 10 to 18  0.7063 -0.6719 -0.6243 -0.2672  0.4823 -0.1681  0.0753  0.1523 -0.6091
#>   0.5739 -0.5357 -0.2301 -0.3829  0.5177 -0.0145 -0.1892  0.1921 -0.1284
#>  -0.3009 -0.2579 -0.0593 -0.0076 -0.1720  0.3191  0.0325 -0.3017  0.3966
#> 
#> Columns 19 to 20 -0.2624  0.3221
#>  -0.3210  0.4765
#>  -0.5340  0.0864
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>