RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0621 -0.7652  0.3280 -0.2556 -0.0907  0.4556 -0.4516  0.3488  0.7675
#>   0.1358  0.6934 -0.2542  0.4017  0.9533 -0.1954 -0.1691 -0.4157 -0.8032
#>   0.1947  0.2059  0.5334  0.3126  0.6886  0.3598  0.1431 -0.3011 -0.2534
#> 
#> Columns 10 to 18  0.0999  0.1205 -0.8120  0.6367 -0.8230 -0.6253  0.4792 -0.3924 -0.5769
#>  -0.9807  0.7295 -0.8606 -0.3424 -0.8259 -0.0604 -0.4281 -0.8774 -0.6969
#>  -0.6855  0.6368 -0.2455  0.2306 -0.0292  0.1776  0.4578  0.0811 -0.4913
#> 
#> Columns 19 to 20 -0.7298  0.1157
#>   0.2831  0.8106
#>  -0.2576  0.0828
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2765 -0.2586 -0.5212 -0.0912  0.2813  0.0586 -0.5063  0.4549 -0.0469
#>   0.2620 -0.3843 -0.3921  0.6463  0.2567 -0.0275 -0.2233 -0.1184  0.3698
#>  -0.5988  0.1937 -0.3711 -0.0469  0.6228 -0.2212 -0.3835  0.1782  0.0150
#> 
#> Columns 10 to 18 -0.0045 -0.0455  0.0060  0.6575  0.0291  0.4260 -0.1684 -0.2318 -0.1809
#>  -0.1566  0.3083 -0.0938  0.6968 -0.2545 -0.0368  0.5257 -0.5239 -0.3035
#>  -0.0365  0.6906  0.2624  0.1691  0.1196 -0.1911 -0.0043 -0.2397 -0.2059
#> 
#> Columns 19 to 20 -0.2170  0.3072
#>  -0.3661  0.4571
#>  -0.1867  0.1807
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0846  0.0532 -0.6242  0.3751  0.6304  0.3936  0.2375  0.2061  0.0810
#>  -0.4302 -0.0488 -0.4435 -0.2388  0.6808  0.0136 -0.5080  0.2375  0.3545
#>   0.3961 -0.3073 -0.0925 -0.2006  0.4148 -0.1301  0.2891  0.1856  0.5722
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3190 -0.0206  0.1833  0.6319 -0.1072  0.3699  0.1032  0.2825 -0.1602
#>  -0.3137  0.6087 -0.2889  0.4501  0.5205 -0.3633  0.4797 -0.3149  0.3170
#>   0.5004 -0.1394 -0.7501  0.1417  0.7557 -0.6204  0.7989  0.0342  0.2288
#> 
#> Columns 10 to 18 -0.3273 -0.3307  0.1142  0.3310  0.0999 -0.5184 -0.2416 -0.4894 -0.3558
#>  -0.7280  0.0069 -0.1031  0.1933  0.1420  0.6800  0.6349 -0.0694 -0.3888
#>  -0.5136 -0.0820  0.5896  0.2370 -0.4142  0.4869  0.4469 -0.2131 -0.2775
#> 
#> Columns 19 to 20 -0.3001  0.1486
#>  -0.2263 -0.3776
#>  -0.4939  0.1624
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3585 -0.2517 -0.3039  0.1292  0.7767  0.0890  0.3416 -0.0420 -0.2656
#>  -0.3938  0.1060 -0.3620 -0.2243  0.5452 -0.3786  0.1621  0.5512 -0.0236
#>  -0.4746  0.2083 -0.4440  0.0730  0.6178  0.0611  0.6360  0.0936 -0.3157
#> 
#> Columns 10 to 18 -0.1401  0.3486 -0.1563  0.3829  0.2769 -0.1615 -0.4860 -0.0213  0.0026
#>   0.1825  0.1885  0.5164  0.5607  0.0841  0.0360 -0.5640 -0.1262  0.4697
#>  -0.2125  0.3837  0.3698  0.2124  0.5472 -0.1666 -0.6705  0.2538  0.6383
#> 
#> Columns 19 to 20 -0.4109  0.2411
#>  -0.2129  0.3651
#>  -0.3030  0.2892
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>