RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1106  0.9196  0.3682 -0.1020  0.3576 -0.5152 -0.7336  0.1150  0.1984
#>   0.4503 -0.3112 -0.0179 -0.0622  0.4143 -0.6737 -0.3603  0.5405 -0.1027
#>  -0.4858  0.3596  0.3498  0.4526  0.7494 -0.5808 -0.4505 -0.3703 -0.2804
#> 
#> Columns 10 to 18 -0.6047  0.4419 -0.8509  0.8060  0.4491  0.7330 -0.8393  0.5810  0.3231
#>   0.3633  0.5800 -0.1590 -0.3731  0.1877  0.6628  0.5920 -0.2934 -0.5961
#>  -0.6698 -0.0099  0.0765  0.8835  0.1328  0.5100  0.2847  0.5299  0.1088
#> 
#> Columns 19 to 20 -0.2687 -0.2537
#>   0.2802 -0.6019
#>  -0.1904  0.2687
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0583  0.0955  0.5634  0.1834 -0.1682 -0.6516 -0.4796  0.0328  0.2122
#>   0.3216 -0.4465 -0.0141  0.0698  0.6940 -0.4306 -0.2432  0.5069 -0.4790
#>  -0.1385  0.2516  0.5498  0.0363 -0.0140 -0.0777 -0.4689 -0.0662  0.3173
#> 
#> Columns 10 to 18  0.3030 -0.0557 -0.2537  0.6028 -0.1887  0.2068 -0.0359  0.2659  0.3554
#>  -0.4169 -0.1243 -0.2495  0.2139 -0.3832 -0.0869  0.4943  0.0777  0.0024
#>   0.4853 -0.1301 -0.4838  0.4697  0.0869 -0.1320 -0.4362  0.2938  0.2867
#> 
#> Columns 19 to 20  0.3757  0.4387
#>   0.0234 -0.0544
#>   0.4733  0.2590
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0639  0.5909  0.5779  0.0364 -0.0782 -0.5052 -0.2480  0.2123  0.4067
#>  -0.0261  0.4193  0.6204 -0.5367  0.4777 -0.4071 -0.6548  0.1383 -0.4111
#>   0.1222 -0.0142  0.3263  0.1885  0.2100 -0.1713 -0.4687  0.5473  0.2420
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3874  0.3190 -0.5374  0.1055  0.1198 -0.0057  0.3595  0.2011 -0.3597
#>   0.5616 -0.5199 -0.8248 -0.2629 -0.4869 -0.3542 -0.1998  0.2411 -0.1608
#>   0.7318  0.3228 -0.4349  0.4960  0.0326 -0.0631 -0.3942 -0.5857  0.2166
#> 
#> Columns 10 to 18 -0.1455 -0.5816  0.2531 -0.7441  0.0678  0.3810  0.3352 -0.4044  0.7942
#>  -0.4152 -0.6826 -0.5162  0.0614 -0.1881  0.2656  0.2611  0.1455 -0.0898
#>  -0.3836 -0.0800 -0.0848 -0.7827  0.4910  0.6682 -0.4919 -0.5194  0.6682
#> 
#> Columns 19 to 20  0.1323 -0.7579
#>   0.2946 -0.2323
#>   0.3236  0.1954
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2961  0.0483  0.5711  0.1699  0.3597 -0.2523 -0.5786  0.3999 -0.1983
#>  -0.1985 -0.2210  0.4621  0.1415  0.7440 -0.2525 -0.2570  0.6517 -0.4802
#>   0.3498 -0.4380  0.4441  0.4629  0.5892 -0.5676 -0.6919  0.8188 -0.6236
#> 
#> Columns 10 to 18  0.0216  0.3003 -0.1479 -0.1383  0.2545 -0.3464  0.2707  0.0819  0.2173
#>  -0.2012 -0.1197 -0.3599  0.2108  0.0530 -0.1312 -0.0029  0.3029  0.1047
#>  -0.2034  0.3739 -0.4892  0.0508  0.0078  0.2007  0.1492 -0.2152 -0.0744
#> 
#> Columns 19 to 20 -0.2312 -0.2626
#>  -0.0736 -0.4453
#>  -0.0887 -0.5421
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>