RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2855 -0.0023  0.4022  0.5913 -0.3378 -0.0322 -0.6338 -0.0854 -0.5575
#>  -0.7790  0.3994 -0.1354  0.2946 -0.2979  0.3686 -0.4300  0.3099 -0.0528
#>  -0.0661 -0.5242  0.6740 -0.0995 -0.5060  0.1510 -0.6897 -0.4707  0.0727
#> 
#> Columns 10 to 18 -0.6041 -0.6788  0.2315 -0.3282  0.5038 -0.8380 -0.0217 -0.5355  0.0964
#>   0.3102 -0.2081  0.0809 -0.6417 -0.7274 -0.8173 -0.4309 -0.6447 -0.3619
#>   0.1068 -0.2463 -0.5832  0.3018  0.2245 -0.5997 -0.5895 -0.5596 -0.7347
#> 
#> Columns 19 to 20  0.5604 -0.0584
#>   0.6076 -0.1051
#>   0.8402 -0.3446
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3336 -0.1260 -0.0306  0.5141  0.1477 -0.1357  0.2722  0.4103  0.6179
#>  -0.5044  0.3274 -0.2101  0.0149 -0.2452  0.4472  0.0109  0.2555 -0.1097
#>   0.3650  0.3400 -0.0318 -0.1847 -0.0484  0.0751 -0.0771 -0.6172 -0.1965
#> 
#> Columns 10 to 18 -0.2930 -0.1275  0.0268  0.1472 -0.3097  0.1295  0.5078 -0.2911 -0.5034
#>  -0.0606 -0.5829 -0.4297  0.5375  0.3581 -0.6912  0.2712  0.0564 -0.0052
#>   0.1688 -0.5362 -0.3273  0.1737  0.0544 -0.4092  0.3732  0.3501 -0.1517
#> 
#> Columns 19 to 20 -0.0155 -0.0180
#>   0.6100 -0.1840
#>   0.3972 -0.1259
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0374 -0.0579  0.2646 -0.1827 -0.4327 -0.4909 -0.5936 -0.6584 -0.1083
#>  -0.3146  0.1543  0.1556 -0.0844 -0.3096 -0.1042 -0.2798 -0.3824 -0.2851
#>  -0.0316 -0.1373  0.3572 -0.0872 -0.2972  0.1778 -0.1744 -0.2116 -0.2646
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2288  0.5038  0.5327 -0.0821 -0.2313 -0.1680  0.3792  0.5031 -0.5821
#>   0.3182 -0.2535  0.7778  0.1537 -0.0060  0.5514 -0.8122  0.0513  0.5490
#>   0.0565 -0.2781  0.2379  0.0430 -0.7623  0.4496 -0.2571  0.6152  0.5590
#> 
#> Columns 10 to 18  0.3099  0.4948 -0.1932  0.4543 -0.2397 -0.7224 -0.1071 -0.2205 -0.7535
#>   0.6205 -0.3651 -0.8466  0.7806 -0.6462 -0.8653  0.2365 -0.1506 -0.6942
#>   0.5761 -0.5449 -0.8029  0.8153  0.7613 -0.5360 -0.2871 -0.0343  0.1597
#> 
#> Columns 19 to 20  0.5189  0.0210
#>  -0.3877  0.5741
#>  -0.3804 -0.5769
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0323 -0.0389 -0.0511  0.3568  0.2083 -0.5597 -0.0999 -0.3439  0.1229
#>  -0.0557  0.7160  0.1468 -0.0570 -0.2156  0.0181 -0.1988 -0.5001 -0.2609
#>  -0.4002  0.4349  0.1114  0.4564  0.0724  0.1223 -0.0197  0.0576  0.1540
#> 
#> Columns 10 to 18 -0.4493 -0.5271  0.1738  0.3374 -0.3370 -0.5406  0.0010 -0.4126 -0.5276
#>  -0.4951 -0.5818 -0.2170 -0.0812 -0.1353 -0.6514  0.3274 -0.4212  0.0878
#>  -0.4953 -0.4415 -0.2788  0.0025  0.2688 -0.6789  0.4175 -0.4016  0.1355
#> 
#> Columns 19 to 20  0.3493 -0.0092
#>   0.6509  0.3296
#>   0.5166 -0.0717
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>