RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5514 -0.4773  0.1233  0.0641  0.5861  0.2990  0.0793  0.2746 -0.6335
#>  -0.2968 -0.0668  0.5888 -0.3042  0.5916 -0.4932 -0.5593 -0.2197 -0.8859
#>   0.7131 -0.1478 -0.8351  0.3351  0.1621 -0.2358 -0.3856 -0.8673 -0.3951
#> 
#> Columns 10 to 18  0.1316 -0.0791 -0.1407  0.2988 -0.2195 -0.3986  0.0916 -0.2231  0.2786
#>  -0.1534 -0.0474  0.9183  0.9598  0.1833 -0.7339 -0.0291 -0.3008 -0.7891
#>  -0.2379  0.9130  0.1902  0.7963 -0.6054  0.0199 -0.2726 -0.7296 -0.4386
#> 
#> Columns 19 to 20 -0.0656 -0.5039
#>   0.7365 -0.7348
#>   0.9729  0.1311
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0756 -0.4849  0.0734 -0.2997  0.5389  0.5785 -0.4544  0.3495 -0.1836
#>   0.2044 -0.2641  0.0266 -0.2122  0.6031  0.7390  0.1447 -0.1916  0.6811
#>  -0.0593  0.2902 -0.3211 -0.6307  0.2263  0.4768  0.6502 -0.7649  0.2801
#> 
#> Columns 10 to 18 -0.0265  0.5582  0.3763  0.2628 -0.2929 -0.3240  0.0207 -0.2132 -0.3298
#>  -0.4938  0.1124  0.5613  0.0202 -0.5936  0.0649 -0.0388 -0.2625  0.1743
#>  -0.2218 -0.4933  0.5606 -0.0064 -0.5382  0.2580  0.3299  0.0659 -0.3267
#> 
#> Columns 19 to 20  0.3610  0.1174
#>  -0.0205 -0.3319
#>   0.0761 -0.2528
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1029  0.5161 -0.0771 -0.0999  0.2866  0.5929  0.5726 -0.0702  0.1363
#>  -0.2298  0.5209  0.5439 -0.1032 -0.0776 -0.0189  0.6656 -0.0836 -0.6055
#>  -0.2650  0.3264  0.1829 -0.1943 -0.1173  0.5284  0.2576 -0.3936 -0.1787
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3165  0.6538 -0.2087  0.2263  0.5133  0.4096  0.3582  0.1050  0.6358
#>  -0.1384  0.4038  0.1897  0.3588  0.6420 -0.4892  0.4375  0.4089 -0.2182
#>  -0.1098  0.1061  0.0591  0.7345 -0.0364 -0.6048  0.2530 -0.2244 -0.7599
#> 
#> Columns 10 to 18  0.1923  0.1757 -0.4645  0.3546 -0.3412 -0.1036 -0.7460  0.7296 -0.5084
#>   0.6889 -0.5404  0.2566  0.0505  0.1995  0.5924 -0.4332  0.3893  0.0155
#>   0.4067 -0.2591  0.1212 -0.0262  0.1120  0.2590 -0.2552  0.7410  0.7953
#> 
#> Columns 19 to 20  0.3327  0.3394
#>   0.1680 -0.2923
#>   0.0266 -0.0991
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4470  0.4653  0.3026  0.1820  0.2240  0.3401  0.3650  0.2098  0.1181
#>  -0.2668 -0.2427  0.2954 -0.0482  0.3191  0.4096 -0.0081  0.4334  0.1757
#>   0.0100 -0.2491  0.2264 -0.0705  0.1900  0.5754  0.3739  0.2807  0.2960
#> 
#> Columns 10 to 18  0.0746  0.0433 -0.1633  0.3242  0.1432 -0.3350  0.1225 -0.2208  0.2525
#>  -0.2164  0.3461  0.0630  0.4584  0.2335 -0.5733  0.5624 -0.3973  0.1835
#>  -0.4659  0.0468  0.2178  0.4037  0.1308 -0.5751  0.5807 -0.5933  0.2230
#> 
#> Columns 19 to 20  0.5396  0.1882
#>   0.2478 -0.0960
#>   0.2484 -0.2580
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>