RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0970 -0.8198  0.4610  0.7394  0.5433  0.3895 -0.3939  0.6503  0.0305
#>   0.0877  0.5476  0.0900  0.3937 -0.0967  0.0772 -0.5019  0.5032 -0.5276
#>  -0.5642  0.3286  0.3468  0.4634 -0.1480  0.3954 -0.5257  0.3562  0.0281
#> 
#> Columns 10 to 18  0.3507  0.0454  0.5561  0.5373  0.7825  0.9254 -0.2150 -0.6169 -0.4274
#>  -0.2535  0.1852  0.1695  0.4649  0.4320  0.2782 -0.5040  0.0870  0.3505
#>   0.4506  0.7099  0.2628  0.2995  0.7166  0.8804 -0.0982 -0.0061 -0.3475
#> 
#> Columns 19 to 20 -0.7537 -0.4173
#>  -0.2099  0.3825
#>  -0.6790  0.0278
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1131 -0.3165 -0.1302  0.1984  0.0290  0.1789 -0.5633 -0.0700  0.4057
#>  -0.1766  0.4385  0.2565 -0.4555  0.4605  0.1666  0.0767 -0.2866  0.6331
#>   0.1291  0.2831 -0.3625  0.4927  0.0567 -0.0715 -0.2803 -0.0653  0.7444
#> 
#> Columns 10 to 18 -0.1629  0.2807  0.1763  0.0784 -0.0374  0.1617  0.2884 -0.6568 -0.3462
#>  -0.3064  0.5472  0.2192  0.3251 -0.1910 -0.3743 -0.0479 -0.0403 -0.3084
#>  -0.0361  0.5685  0.2809 -0.3746 -0.3222  0.3385  0.1481 -0.4893  0.0068
#> 
#> Columns 19 to 20  0.0820 -0.2628
#>   0.3060  0.0542
#>  -0.3102 -0.2325
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2920  0.2078  0.3348 -0.1807  0.0500  0.2702 -0.4010 -0.1426  0.2204
#>  -0.7039  0.3080  0.4572  0.3726 -0.5210  0.1965  0.1662 -0.0696  0.0122
#>  -0.0150 -0.1526  0.1737  0.3728 -0.1034 -0.2251 -0.3219 -0.6881  0.2041
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0017 -0.1638 -0.1126 -0.4507 -0.8067  0.8435 -0.4047  0.5840  0.2240
#>   0.3215 -0.2119 -0.5856  0.2398 -0.3899  0.1118 -0.4292 -0.3604 -0.0754
#>   0.7283  0.6109 -0.4913  0.3960 -0.3713 -0.0157 -0.0317 -0.3022 -0.0242
#> 
#> Columns 10 to 18  0.0831  0.0201  0.4751  0.6253  0.2328 -0.1622 -0.4649  0.3210 -0.5097
#>   0.2724  0.4748 -0.2323  0.5537  0.1314 -0.8210  0.1782 -0.2496 -0.0621
#>  -0.1685  0.3386 -0.4148  0.2683 -0.1905 -0.4183  0.5035 -0.3787 -0.0194
#> 
#> Columns 19 to 20  0.1167 -0.2547
#>   0.4457 -0.8241
#>   0.4323 -0.5373
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5453  0.2017  0.3644  0.3045 -0.0438  0.3275  0.1439 -0.1617  0.1566
#>  -0.3091  0.2732  0.0505 -0.0084  0.1758  0.2909  0.3728  0.3342 -0.4242
#>  -0.3497  0.3585 -0.0526  0.0242  0.1751  0.1334  0.3502  0.2049 -0.4490
#> 
#> Columns 10 to 18  0.2837  0.4538  0.0228  0.3178 -0.0391  0.7247 -0.5445  0.3104 -0.3932
#>   0.6387  0.2646  0.3143  0.4346  0.0224  0.3712 -0.2422  0.5868 -0.4434
#>   0.6918  0.4227  0.4215  0.2382 -0.1004 -0.0048 -0.2413  0.3811 -0.4407
#> 
#> Columns 19 to 20  0.4684 -0.2108
#>   0.0389 -0.3815
#>   0.0245 -0.5618
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>