RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6253 -0.3089  0.3210 -0.5107 -0.1109  0.0855 -0.2510  0.2457  0.4549
#>   0.6142 -0.0126 -0.1603 -0.7392  0.8341 -0.2209  0.5939  0.2997 -0.6605
#>  -0.4555 -0.2131 -0.5352  0.0310  0.3037  0.2139 -0.2350 -0.0458 -0.1334
#> 
#> Columns 10 to 18 -0.1166 -0.4472  0.0581  0.6492  0.3274  0.6914  0.7343 -0.4623  0.5061
#>  -0.6683 -0.6559 -0.5248  0.6659 -0.0419  0.2265  0.0615 -0.2509 -0.3841
#>   0.0733  0.7977 -0.2267 -0.3407 -0.3458 -0.2400  0.3916 -0.3217  0.3686
#> 
#> Columns 19 to 20  0.1206  0.0801
#>   0.3041 -0.3688
#>   0.4405 -0.4299
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2001 -0.4861 -0.1332  0.0869  0.2812 -0.4609  0.4261 -0.0124 -0.0606
#>   0.5929 -0.7821  0.0938 -0.1404  0.4919 -0.6481  0.4279  0.4961 -0.3618
#>  -0.0667 -0.1745 -0.4709  0.1744  0.4203 -0.0688  0.4482  0.5728 -0.1399
#> 
#> Columns 10 to 18 -0.2011  0.1360  0.0747 -0.0822  0.3490 -0.0172  0.1406  0.3340 -0.3976
#>   0.0572 -0.0937  0.0036  0.4309 -0.1604  0.3754 -0.0536  0.3795 -0.3604
#>  -0.2158  0.0193 -0.2530 -0.4098  0.3555 -0.0706  0.5558 -0.0492  0.2199
#> 
#> Columns 19 to 20 -0.5194  0.0195
#>  -0.4205 -0.5257
#>  -0.1794 -0.6557
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4067 -0.3459 -0.1785  0.2472  0.3773  0.0730  0.1493  0.1335 -0.0746
#>   0.3821 -0.5520  0.0303  0.1153  0.5443 -0.0424  0.1632 -0.1777 -0.0163
#>   0.4821 -0.4534 -0.2470  0.3964  0.3786 -0.0768  0.3604  0.1554  0.0130
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4264 -0.0060 -0.2904 -0.4317 -0.7311 -0.2628  0.2259 -0.8724 -0.3519
#>   0.2299  0.4203  0.8660 -0.3853 -0.0441 -0.0912 -0.6898 -0.2983  0.6881
#>  -0.2284 -0.0749  0.0108 -0.2741  0.1051  0.3295 -0.1108 -0.2865 -0.1652
#> 
#> Columns 10 to 18 -0.4196  0.4276 -0.6068 -0.3357 -0.2424 -0.3159 -0.1884 -0.7383 -0.0153
#>  -0.4226  0.5376  0.6336  0.1174 -0.6973 -0.6057  0.4649  0.2908  0.0508
#>   0.0348  0.3599 -0.0573 -0.0424  0.1480  0.2427 -0.5094 -0.5742 -0.0058
#> 
#> Columns 19 to 20 -0.6795  0.5418
#>  -0.3718  0.8673
#>  -0.2598  0.4021
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2354 -0.6096 -0.4551  0.2445  0.5534  0.0548  0.4846  0.4360 -0.4482
#>   0.5307 -0.6107  0.2837 -0.1091  0.5720 -0.6870  0.5485  0.0689 -0.3945
#>   0.3244 -0.7034 -0.3164 -0.0399  0.3509 -0.2557  0.3583 -0.1725 -0.2983
#> 
#> Columns 10 to 18 -0.3416 -0.0955 -0.2256  0.0820  0.0014 -0.1005  0.3047  0.2563 -0.2084
#>  -0.6201 -0.3174 -0.1569  0.6137  0.0592  0.2938  0.0985  0.0876 -0.4052
#>  -0.0554 -0.1933 -0.1914  0.4268 -0.1869  0.0777  0.0744  0.0568  0.1006
#> 
#> Columns 19 to 20 -0.2040 -0.5255
#>  -0.5326  0.3292
#>  -0.3568 -0.3046
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>