RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8414 -0.4090  0.1463  0.6082 -0.4356  0.3193 -0.2034  0.0958 -0.2693
#>   0.1396  0.2818 -0.4271 -0.7584 -0.8665 -0.8750  0.0492 -0.5716 -0.2169
#>   0.7923 -0.2672  0.7407 -0.2291  0.0966  0.3918 -0.1006 -0.4503 -0.6783
#> 
#> Columns 10 to 18  0.9438  0.6578  0.6927  0.2947 -0.0492  0.3942 -0.1222  0.7902  0.1771
#>   0.6304  0.0401 -0.0025  0.6201  0.0900 -0.1960  0.5256  0.4820  0.4594
#>  -0.7070 -0.3555  0.1093  0.1022  0.2656  0.5781 -0.5117 -0.9471  0.5513
#> 
#> Columns 19 to 20  0.5714  0.6481
#>  -0.6489  0.7632
#>   0.5029 -0.4914
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6141 -0.3544  0.3501 -0.5783 -0.4112 -0.2734 -0.1611 -0.0160  0.0024
#>  -0.0046  0.4577 -0.2495  0.4006 -0.6649  0.0140  0.5026 -0.1432  0.2838
#>   0.3318  0.3295  0.1203 -0.0079 -0.1129  0.3465 -0.6479  0.0237 -0.5044
#> 
#> Columns 10 to 18  0.3868 -0.4821  0.0187 -0.6255  0.4211  0.0130  0.4678  0.4552  0.2794
#>   0.2716  0.0607  0.2300 -0.1098  0.0984 -0.2361 -0.2561  0.7034 -0.1932
#>   0.1412  0.5413  0.4453  0.0180  0.4420 -0.2194  0.1562  0.2982  0.2118
#> 
#> Columns 19 to 20 -0.6234  0.6439
#>   0.0400  0.5498
#>  -0.2501 -0.1783
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0112  0.5532  0.7537 -0.2838  0.1601  0.3215 -0.4506 -0.2941 -0.2793
#>   0.0372 -0.3481  0.4976 -0.2472 -0.6212 -0.0235 -0.5343  0.0375 -0.5130
#>   0.0167 -0.1709 -0.0864  0.2958 -0.7113 -0.3080 -0.1752  0.0289  0.0370
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1623 -0.6090  0.8628 -0.6249  0.8530 -0.5353  0.3894 -0.1424  0.4234
#>  -0.6053  0.3721  0.1506  0.1247  0.8252 -0.5287 -0.0522 -0.1236 -0.2627
#>  -0.4165 -0.4219 -0.5055 -0.1433  0.4170  0.0277  0.1317 -0.0251 -0.0629
#> 
#> Columns 10 to 18  0.7410  0.6096 -0.8282 -0.0621 -0.0998 -0.4982  0.2517  0.1394  0.6520
#>   0.2903  0.0414 -0.3641 -0.8148 -0.7808 -0.3369  0.1500 -0.4792 -0.4676
#>  -0.6229  0.1528  0.0941 -0.1014  0.1247 -0.2780  0.3775 -0.2282  0.3113
#> 
#> Columns 19 to 20 -0.6270 -0.0978
#>  -0.4821 -0.9631
#>  -0.3575  0.4347
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5404  0.0961  0.5331 -0.0999 -0.3205  0.1999 -0.3335 -0.1289 -0.2606
#>   0.2079  0.0519  0.0243 -0.6006 -0.3651  0.2991 -0.2443 -0.1946 -0.4884
#>  -0.0099 -0.2599  0.1443 -0.2997 -0.5818 -0.1842 -0.4852  0.1367 -0.1127
#> 
#> Columns 10 to 18 -0.0974 -0.3218  0.7616  0.0941  0.3472  0.1680  0.3256  0.2535  0.0919
#>  -0.3400 -0.6641  0.5973  0.1545  0.4101 -0.2013  0.2587  0.4031 -0.0009
#>   0.1554 -0.3806  0.1034 -0.3332  0.2857 -0.0144  0.1912  0.3142 -0.0475
#> 
#> Columns 19 to 20 -0.2283  0.6374
#>  -0.4662  0.7167
#>  -0.3221  0.3799
#> [ CPUFloatType{2,3,20} ]
#>