RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.7161  0.2524 -0.2267 -0.2711 -0.2845 -0.4067  0.5843 -0.1285 -0.5132
#>   0.6516  0.7333  0.6177  0.2089 -0.5074  0.5751  0.5784  0.6827  0.0714
#>   0.2538  0.6949 -0.0763  0.5675 -0.3406 -0.1142  0.4442 -0.0423 -0.0378
#> 
#> Columns 10 to 18 -0.1102  0.6036  0.1749 -0.0608 -0.4173 -0.0102 -0.2940  0.4289 -0.1320
#>  -0.3589 -0.2381  0.0580  0.7796  0.0998 -0.2888 -0.1735 -0.8531 -0.4652
#>   0.1831  0.0378 -0.6822 -0.2816  0.1889 -0.2580 -0.0010 -0.2612  0.1988
#> 
#> Columns 19 to 20  0.3093 -0.4167
#>   0.0837  0.0514
#>   0.7636  0.3214
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5985  0.0873  0.5781 -0.1434 -0.1528  0.1490  0.2079  0.7476 -0.1045
#>   0.3875  0.0819  0.5025  0.3552 -0.2228  0.5654 -0.0206  0.6165 -0.0171
#>   0.4864  0.6804  0.0500  0.3966 -0.0198 -0.0722 -0.0076  0.4736 -0.2741
#> 
#> Columns 10 to 18 -0.1348  0.2238  0.0753 -0.3945  0.1603  0.2999  0.2501 -0.5067 -0.6350
#>  -0.3108 -0.3115 -0.2025  0.3937  0.4892 -0.2319 -0.2614 -0.6291  0.0740
#>  -0.5959 -0.0322 -0.2894  0.1512  0.3782 -0.1010  0.1482 -0.3825  0.4464
#> 
#> Columns 19 to 20  0.2921 -0.1500
#>   0.6855 -0.0422
#>   0.9088 -0.0266
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.6381  0.1431  0.4164  0.3592  0.0609  0.3689  0.1772  0.4760  0.0470
#>   0.4289 -0.1033  0.4358  0.1910 -0.0969  0.5367  0.1665  0.4100  0.0255
#>   0.2304  0.1687  0.5151  0.1998  0.2694  0.3913  0.2137 -0.1146 -0.1561
#> 
#> Columns 10 to 18 -0.4619 -0.2263 -0.2680  0.3368  0.2228  0.1836  0.2800 -0.6207 -0.2572
#>  -0.1175 -0.2322 -0.6100  0.0511  0.2566 -0.2520 -0.2023 -0.6280 -0.1511
#>  -0.3149 -0.2550 -0.2557  0.1514 -0.2146  0.2289  0.2341 -0.7046 -0.1510
#> 
#> Columns 19 to 20  0.4382  0.3621
#>   0.4396  0.0057
#>   0.6035 -0.3589
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.2884 -0.0291  0.1229  0.0041 -0.1166  0.3444  0.0175  0.3235  0.0658
#>   0.6147  0.4537  0.1856  0.1567 -0.3019  0.1505  0.2564  0.5269  0.1842
#>   0.4272  0.2199  0.5595  0.2342 -0.1921 -0.1427  0.3725  0.0805  0.0195
#> 
#> Columns 10 to 18 -0.0812 -0.0068 -0.1565 -0.0358 -0.0224 -0.2966 -0.2252 -0.4043  0.0966
#>  -0.3598 -0.1854 -0.3976  0.0953  0.1748 -0.5111  0.1526 -0.4405 -0.2781
#>  -0.3753  0.0663 -0.1238 -0.0727 -0.0397 -0.0635  0.0157 -0.0993 -0.1252
#> 
#> Columns 19 to 20  0.5735 -0.1105
#>   0.7309  0.1156
#>   0.7796  0.3005
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.6705  0.3787  0.4285  0.3407  0.0861  0.0279  0.3407  0.3227 -0.1617
#>   0.5895  0.1572  0.6036  0.4525  0.0566  0.3992  0.5467  0.0065 -0.2110
#>   0.1297 -0.0524  0.5491  0.2352  0.2431  0.4329 -0.1639  0.5197 -0.1087
#> 
#> Columns 10 to 18 -0.1388 -0.0327 -0.5949  0.0944 -0.1133 -0.1225  0.1684 -0.6087 -0.1145
#>  -0.3481 -0.1070 -0.5201  0.3679  0.1345  0.1734  0.1243 -0.5208  0.0556
#>  -0.4741 -0.3434 -0.3230 -0.2729  0.3473  0.0649  0.2237 -0.2233 -0.0649
#> 
#> Columns 19 to 20  0.6030  0.2479
#>   0.6858 -0.1467
#>   0.6513 -0.2089
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.0992 -0.6436 -0.3197  0.6220  0.2256 -0.7712  0.2246 -0.1414  0.3241
#>   0.3420 -0.6408 -0.3408  0.5502 -0.2351 -0.0798  0.3823 -0.7235  0.1726
#>  -0.1093 -0.3336 -0.2989  0.4091  0.1061  0.5667 -0.4299 -0.3630  0.3574
#> 
#> Columns 10 to 18  0.2762 -0.1739 -0.1027  0.3692 -0.6185 -0.1532 -0.2170  0.0237 -0.4784
#>   0.6071 -0.0503 -0.7492  0.4626 -0.3258  0.1729 -0.4063  0.0940 -0.1114
#>   0.0344 -0.3348  0.5308 -0.3231  0.3965 -0.0367 -0.0665 -0.2350  0.4546
#> 
#> Columns 19 to 20  0.5460 -0.0594
#>   0.1078  0.4941
#>  -0.2752  0.2571
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6705  0.3787  0.4285  0.3407  0.0861  0.0279  0.3407  0.3227 -0.1617
#>   0.5895  0.1572  0.6036  0.4525  0.0566  0.3992  0.5467  0.0065 -0.2110
#>   0.1297 -0.0524  0.5491  0.2352  0.2431  0.4329 -0.1639  0.5197 -0.1087
#> 
#> Columns 10 to 18 -0.1388 -0.0327 -0.5949  0.0944 -0.1133 -0.1225  0.1684 -0.6087 -0.1145
#>  -0.3481 -0.1070 -0.5201  0.3679  0.1345  0.1734  0.1243 -0.5208  0.0556
#>  -0.4741 -0.3434 -0.3230 -0.2729  0.3473  0.0649  0.2237 -0.2233 -0.0649
#> 
#> Columns 19 to 20  0.6030  0.2479
#>   0.6858 -0.1467
#>   0.6513 -0.2089
#> [ CPUFloatType{2,3,20} ]
#>