RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\text{ReLU}\) non-linearity to an input sequence.

nn_rnn(...)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional:
If TRUE, becomes a bidirectional RNN. Default: FALSE

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\text{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\text{input_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\text{hidden_size}\) Defaults to zero if not provided. where \(S=\text{num_layers} * \text{num_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\text{num_directions} * \text{hidden_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_lk: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_lk: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_lk: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_lk: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\text{hidden\_size}}\)

Examples

rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7707 -0.2127  0.0833 -0.5020  0.7743  0.9150  0.0502 -0.3490  0.9177
#>  -0.7058 -0.3404  0.9008  0.5612  0.1238  0.1118  0.2086  0.3772 -0.2031
#>  -0.3612 -0.1235 -0.7028  0.4200  0.0411 -0.2072  0.5555  0.9468 -0.0299
#> 
#> Columns 10 to 18  0.1455  0.6700  0.8668  0.7365  0.5706  0.5940  0.5845 -0.3978 -0.3471
#>   0.0042 -0.0642  0.0454 -0.2421 -0.6776  0.0689 -0.4781  0.5221 -0.8138
#>  -0.7804  0.6880 -0.6256  0.5869 -0.5678 -0.0954  0.0748 -0.7484  0.2925
#> 
#> Columns 19 to 20  0.2962 -0.4581
#>   0.8697  0.9076
#>   0.5304  0.0117
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3931  0.3354 -0.6032  0.5553 -0.0911 -0.0173 -0.4029  0.1835  0.4575
#>  -0.7199 -0.4652  0.2671 -0.4557  0.0219 -0.0998 -0.2311 -0.6716  0.0719
#>  -0.6873 -0.4190  0.2820  0.4307 -0.3875 -0.0353  0.1192  0.1583  0.0578
#> 
#> Columns 10 to 18  0.0573 -0.3042 -0.2850  0.2343  0.1400 -0.2909  0.7082  0.3566 -0.0579
#>   0.5340 -0.3909 -0.1861 -0.2526  0.1529  0.1250 -0.0907  0.2993  0.2565
#>   0.2468 -0.6759 -0.1813  0.1958 -0.1157  0.7653  0.0484 -0.4860  0.5066
#> 
#> Columns 19 to 20 -0.0106  0.5368
#>   0.6679 -0.0620
#>   0.8322  0.0632
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2025 -0.6474  0.4180 -0.3793 -0.0460  0.2731  0.0927 -0.4733 -0.0691
#>  -0.4081 -0.5279 -0.0483 -0.2613 -0.1064 -0.2496  0.1590 -0.2841  0.4583
#>  -0.5098  0.0026 -0.3198  0.6048 -0.1103  0.2459 -0.0079 -0.2687  0.2318
#> 
#> Columns 10 to 18  0.0581  0.4398 -0.3201  0.3175 -0.0661  0.4878 -0.0683 -0.0018  0.1940
#>   0.0484 -0.1143  0.0090 -0.1148  0.0090  0.2331  0.1753  0.0266  0.1241
#>   0.4600 -0.2172 -0.1378 -0.4170 -0.1131  0.6376  0.5091 -0.1358 -0.2057
#> 
#> Columns 19 to 20  0.3009 -0.2259
#>  -0.0075  0.0121
#>   0.5144  0.1673
#> 
#> (4,.,.) = 
#>  Columns 1 to 9 -0.4531 -0.2168 -0.5408 -0.0071 -0.2930  0.0422 -0.3546 -0.4449  0.3840
#>  -0.3371 -0.4406 -0.0102 -0.1182 -0.3625  0.1465 -0.3650  0.0263  0.5210
#>  -0.3134 -0.1051 -0.2616  0.0008  0.1564  0.3192 -0.0420 -0.0784  0.3283
#> 
#> Columns 10 to 18  0.3663  0.1935 -0.2922 -0.4084  0.1909  0.2754  0.4684 -0.0819  0.2299
#>   0.1765 -0.1040 -0.4538  0.2761  0.1063  0.2704  0.2483 -0.0832 -0.2099
#>   0.0862  0.0529 -0.1402 -0.1801 -0.1077  0.4180  0.2165  0.1804  0.0430
#> 
#> Columns 19 to 20 -0.1043 -0.1730
#>   0.2977  0.5321
#>   0.4780  0.0898
#> 
#> (5,.,.) = 
#>  Columns 1 to 9 -0.1420 -0.4948 -0.0771 -0.4554 -0.1340  0.0152 -0.2039  0.1756  0.5272
#>   0.0046 -0.2156 -0.1057  0.0029 -0.1536  0.0371 -0.0947 -0.3541  0.3468
#>  -0.3493 -0.0818  0.0744  0.2137  0.3389 -0.0741  0.0560  0.0755  0.3899
#> 
#> Columns 10 to 18 -0.1118  0.1548 -0.1897  0.4725  0.3748  0.1778 -0.0191 -0.2158  0.0108
#>   0.1313  0.4106 -0.4301 -0.4310  0.4229  0.5693  0.2579 -0.1599  0.3064
#>   0.2120 -0.2889  0.5255  0.1630 -0.2614  0.4556 -0.1420  0.3162 -0.0570
#> 
#> Columns 19 to 20  0.0304 -0.1577
#>   0.2358 -0.2714
#>   0.2031  0.0557
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6077  0.0478  0.0164 -0.2950 -0.2971 -0.4781  0.0169  0.2555  0.0133
#>  -0.4616  0.1612  0.5310  0.1255 -0.2302  0.0283  0.8916  0.6087 -0.3284
#>  -0.2969 -0.3002 -0.5179  0.6761 -0.2708 -0.5304  0.6671 -0.4906  0.0421
#> 
#> Columns 10 to 18  0.6736  0.2246  0.1710  0.3099 -0.1167  0.0170  0.5478 -0.2716  0.2036
#>   0.6366 -0.3136  0.6336 -0.4472 -0.4417  0.4317  0.2124  0.4570  0.1472
#>  -0.4469 -0.3307 -0.1321 -0.3429  0.3623 -0.3800  0.4302  0.1987 -0.2188
#> 
#> Columns 19 to 20  0.0917  0.2792
#>  -0.0328  0.1105
#>   0.0621 -0.1483
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1420 -0.4948 -0.0771 -0.4554 -0.1340  0.0152 -0.2039  0.1756  0.5272
#>   0.0046 -0.2156 -0.1057  0.0029 -0.1536  0.0371 -0.0947 -0.3541  0.3468
#>  -0.3493 -0.0818  0.0744  0.2137  0.3389 -0.0741  0.0560  0.0755  0.3899
#> 
#> Columns 10 to 18 -0.1118  0.1548 -0.1897  0.4725  0.3748  0.1778 -0.0191 -0.2158  0.0108
#>   0.1313  0.4106 -0.4301 -0.4310  0.4229  0.5693  0.2579 -0.1599  0.3064
#>   0.2120 -0.2889  0.5255  0.1630 -0.2614  0.4556 -0.1420  0.3162 -0.0570
#> 
#> Columns 19 to 20  0.0304 -0.1577
#>   0.2358 -0.2714
#>   0.2031  0.0557
#> [ CPUFloatType{2,3,20} ]
#>