RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\text{ReLU}\) non-linearity to an input sequence.

nn_rnn(...)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional:
If TRUE, becomes a bidirectional RNN. Default: FALSE

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\text{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\text{input_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\text{hidden_size}\) Defaults to zero if not provided. where \(S=\text{num_layers} * \text{num_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\text{num_directions} * \text{hidden_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_lk: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_lk: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_lk: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_lk: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\text{hidden\_size}}\)

Examples

rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.3655 -0.6286 -0.0992  0.5596  0.4560 -0.1168  0.5992  0.4446 -0.6654
#>  -0.9142 -0.5876  0.1058 -0.1802  0.2969  0.6124  0.8718 -0.3705  0.0003
#>  -0.8245 -0.0760  0.0132  0.6433  0.2870 -0.4166  0.7227  0.0784 -0.5689
#> 
#> Columns 10 to 18  0.2411  0.6718  0.1353 -0.8511 -0.1119  0.3238  0.8848 -0.5603  0.2245
#>  -0.9390 -0.0850  0.5162 -0.8345 -0.6790 -0.6383  0.4651 -0.6200 -0.8992
#>   0.5944  0.6580  0.6748  0.0202 -0.6313  0.6461 -0.5871  0.6201 -0.8199
#> 
#> Columns 19 to 20  0.2834  0.3641
#>  -0.0829 -0.9525
#>   0.4886 -0.6176
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4630 -0.1654  0.0366  0.0488  0.0255  0.1397 -0.0467  0.0167 -0.3177
#>   0.0460 -0.4872 -0.2313  0.5883 -0.0541  0.0143  0.4212  0.2199  0.3608
#>   0.6325 -0.6333  0.2637  0.5658  0.2123 -0.0034  0.3100  0.3414  0.1747
#> 
#> Columns 10 to 18  0.5703  0.5156 -0.1281 -0.3136 -0.0477  0.2962 -0.1233  0.0840  0.2240
#>  -0.5812  0.5285 -0.4178 -0.7397 -0.0010 -0.3267  0.6574 -0.6126 -0.4487
#>   0.1596  0.7205 -0.1754 -0.4565  0.1168  0.0596 -0.0099  0.2179  0.5875
#> 
#> Columns 19 to 20  0.1263 -0.2559
#>  -0.5976  0.1905
#>  -0.0776  0.2986
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2602 -0.2552  0.1617  0.3524 -0.0182  0.0431  0.3075  0.4382 -0.5057
#>  -0.3376  0.3770  0.4925  0.1848  0.0575  0.4728 -0.4025  0.1534 -0.1865
#>  -0.4849 -0.1757  0.2727 -0.0249  0.0226  0.1648 -0.2488  0.3024 -0.1325
#> 
#> Columns 10 to 18  0.5435  0.0794  0.5435 -0.3847  0.4734  0.2827  0.1959 -0.5269  0.6122
#>   0.6034  0.1883  0.4192 -0.3645 -0.1440  0.5911  0.3576 -0.7983  0.6369
#>   0.5206  0.0596  0.2219 -0.1890 -0.2760  0.4148 -0.0679 -0.1083  0.1574
#> 
#> Columns 19 to 20 -0.2740  0.0649
#>  -0.4336  0.3151
#>  -0.1529 -0.3369
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.2526 -0.3261 -0.1113 -0.4837 -0.1436  0.0717 -0.0099 -0.1939  0.1929
#>   0.0347 -0.6624 -0.3763  0.2697 -0.1457  0.4346  0.4647 -0.0941 -0.5179
#>   0.0683  0.0886  0.0676  0.0308 -0.0394 -0.4712  0.0044 -0.0340 -0.3483
#> 
#> Columns 10 to 18 -0.0279  0.1275 -0.3935 -0.5656 -0.1414  0.0720 -0.3806  0.2405 -0.3331
#>   0.1140  0.0747  0.4420 -0.4194 -0.1293  0.2425 -0.0226 -0.5316  0.3947
#>   0.4664  0.2869  0.7065 -0.1979  0.3610  0.3093 -0.3078 -0.4691  0.3898
#> 
#> Columns 19 to 20 -0.0172 -0.1281
#>   0.4058 -0.5178
#>  -0.4131  0.4689
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.2817  0.3657  0.5170  0.2708 -0.0723  0.0653  0.0710  0.3355 -0.4124
#>   0.1272 -0.1985  0.2120 -0.0095 -0.2502  0.0465 -0.2054 -0.1133  0.2888
#>   0.3501  0.0123 -0.0372 -0.2039 -0.2996  0.2770 -0.3369  0.2206  0.6675
#> 
#> Columns 10 to 18  0.5122 -0.0692  0.7531 -0.3546  0.4882  0.2418  0.5268 -0.6694  0.3564
#>   0.0357  0.2108 -0.2110 -0.4114  0.2128 -0.2155 -0.2007 -0.0955 -0.0687
#>  -0.2097  0.1193  0.0193 -0.6267 -0.5280  0.1809 -0.0165 -0.3595  0.2879
#> 
#> Columns 19 to 20 -0.6337  0.1795
#>  -0.2797 -0.0056
#>  -0.1882 -0.2314
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.7841  0.1349  0.5161 -0.8821  0.0646  0.1561  0.5637 -0.3626 -0.3491
#>  -0.3782  0.0391 -0.2675  0.1728 -0.0541 -0.0544  0.3597  0.2820 -0.3730
#>  -0.8382  0.5585  0.3529  0.6434 -0.5874 -0.0069 -0.2726  0.0330 -0.7796
#> 
#> Columns 10 to 18 -0.4217  0.8123  0.4195  0.1791  0.6232 -0.8344 -0.1124  0.1184  0.1913
#>  -0.6034 -0.6732 -0.0173 -0.6015 -0.1402  0.4507  0.3818 -0.3360  0.6471
#>  -0.3941 -0.5646  0.1631 -0.9252  0.6348  0.7117  0.1990  0.2826  0.4283
#> 
#> Columns 19 to 20  0.6360  0.0452
#>  -0.2111 -0.1405
#>  -0.4239 -0.5370
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2817  0.3657  0.5170  0.2708 -0.0723  0.0653  0.0710  0.3355 -0.4124
#>   0.1272 -0.1985  0.2120 -0.0095 -0.2502  0.0465 -0.2054 -0.1133  0.2888
#>   0.3501  0.0123 -0.0372 -0.2039 -0.2996  0.2770 -0.3369  0.2206  0.6675
#> 
#> Columns 10 to 18  0.5122 -0.0692  0.7531 -0.3546  0.4882  0.2418  0.5268 -0.6694  0.3564
#>   0.0357  0.2108 -0.2110 -0.4114  0.2128 -0.2155 -0.2007 -0.0955 -0.0687
#>  -0.2097  0.1193  0.0193 -0.6267 -0.5280  0.1809 -0.0165 -0.3595  0.2879
#> 
#> Columns 19 to 20 -0.6337  0.1795
#>  -0.2797 -0.0056
#>  -0.1882 -0.2314
#> [ CPUFloatType{2,3,20} ]
#>