RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8596  0.5979  0.0449  0.3652  0.8610 -0.8559  0.8525  0.8955 -0.0654
#>   0.1968  0.3651  0.2952 -0.8445 -0.3174  0.7412  0.5666  0.7422 -0.0125
#>   0.6934 -0.1668  0.8694 -0.3860 -0.0821  0.7006  0.2308 -0.4817  0.8325
#> 
#> Columns 10 to 18 -0.2092  0.8730 -0.5593  0.7362 -0.7178  0.1129  0.4817  0.5061  0.1243
#>  -0.6160 -0.7496  0.9419  0.8880  0.4249 -0.5903 -0.3675 -0.7930  0.8795
#>  -0.4355  0.6397  0.5087  0.6309  0.4433  0.0879 -0.0480 -0.4662  0.8097
#> 
#> Columns 19 to 20 -0.6208 -0.6243
#>   0.0341  0.8400
#>   0.8752  0.4634
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5310  0.0431  0.4238  0.0205 -0.0801 -0.2023 -0.0339 -0.0464  0.1827
#>   0.2092 -0.7955 -0.4921  0.4644 -0.6352 -0.6336  0.1388  0.0158  0.1417
#>   0.1279 -0.3523  0.6456  0.4475 -0.0625 -0.1952 -0.2551 -0.0138 -0.2572
#> 
#> Columns 10 to 18 -0.5556 -0.1169  0.1311  0.7991 -0.2439 -0.1521  0.6091 -0.6184  0.1283
#>   0.1865  0.1688  0.1860  0.0385 -0.3146  0.5837  0.4816 -0.7321  0.1079
#>   0.0836 -0.1969  0.4967  0.5055 -0.1563  0.6960  0.0903 -0.5842 -0.2932
#> 
#> Columns 19 to 20 -0.2877 -0.3196
#>  -0.6654 -0.5003
#>  -0.4222 -0.4928
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2564 -0.6855 -0.1506  0.2257 -0.1490 -0.7179  0.3449  0.5733 -0.0375
#>  -0.1500  0.2833  0.4216  0.2846  0.3485 -0.0837  0.2449  0.2436  0.3688
#>  -0.3410 -0.5531  0.3185  0.4162  0.0789 -0.5890  0.3941  0.6009  0.2352
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1221 -0.5625 -0.1624 -0.3655  0.3224  0.3104 -0.6915 -0.0046 -0.5672
#>  -0.0491 -0.2155 -0.1702 -0.5771 -0.1809  0.4857 -0.2385 -0.4034  0.2691
#>  -0.5761  0.3042 -0.5014 -0.4894  0.4748  0.6871  0.1269 -0.7355  0.9104
#> 
#> Columns 10 to 18  0.3402 -0.2230  0.0602 -0.1211  0.2572  0.4433  0.6114  0.7161 -0.1003
#>   0.3332  0.5529  0.4022  0.0368 -0.0167 -0.3349 -0.2498  0.3303 -0.3365
#>  -0.8024 -0.5069  0.4985  0.3265 -0.2474  0.9183 -0.3971  0.7436 -0.6305
#> 
#> Columns 19 to 20  0.1269  0.5634
#>  -0.3856 -0.7468
#>  -0.5823 -0.6543
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1083 -0.3152  0.4878  0.3823 -0.0924 -0.5699  0.4784  0.5292  0.2987
#>   0.1404 -0.3911  0.4238  0.4470  0.1107 -0.4463  0.5525  0.6011  0.3858
#>   0.3437 -0.4625  0.5377  0.3000 -0.3235 -0.4995  0.1028  0.5557  0.5257
#> 
#> Columns 10 to 18 -0.3540  0.4703 -0.4820  0.3913 -0.4828  0.2726  0.2555 -0.4463 -0.2045
#>  -0.1079 -0.5323 -0.0612  0.1943 -0.4066 -0.1655  0.6225 -0.4063  0.2735
#>  -0.4929 -0.6246 -0.3836  0.5329  0.1994  0.1921  0.2954 -0.5413  0.3754
#> 
#> Columns 19 to 20 -0.3035 -0.5207
#>  -0.2659 -0.5125
#>  -0.0463 -0.6026
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>