RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7965 -0.2735  0.4606  0.7795  0.3596 -0.6715 -0.3950  0.1430 -0.4538
#>   0.8029  0.2185  0.1503 -0.0481  0.7651  0.5116  0.0160 -0.1962  0.2711
#>  -0.6711 -0.7650 -0.3506 -0.3386 -0.7979  0.2205  0.9185 -0.4885 -0.6062
#> 
#> Columns 10 to 18  0.3010 -0.1124 -0.4075 -0.1402 -0.2919 -0.2439 -0.2362 -0.0577  0.2096
#>  -0.7178 -0.3923 -0.5034  0.2524  0.2125  0.0719  0.8017  0.1144 -0.0350
#>   0.5518 -0.5289  0.8510 -0.6611 -0.4299  0.3363 -0.7290  0.5113  0.1416
#> 
#> Columns 19 to 20  0.1647  0.5771
#>   0.3950  0.7712
#>   0.6479 -0.8410
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0968  0.3073 -0.1838 -0.3421 -0.0348  0.3058  0.2322  0.6648 -0.1238
#>  -0.1716  0.4855 -0.4567 -0.0807 -0.4070  0.0018 -0.1547 -0.4091 -0.5548
#>  -0.4558 -0.4546 -0.2917 -0.0703  0.5610  0.1419  0.5083  0.6692 -0.1716
#> 
#> Columns 10 to 18 -0.4108 -0.6393 -0.3976 -0.0473 -0.1743 -0.2632  0.2730 -0.0384  0.6309
#>  -0.2672 -0.6048  0.2657 -0.6045 -0.0371 -0.4432 -0.1887  0.2173 -0.2761
#>   0.2071  0.0767 -0.2771  0.6869 -0.1978  0.0944  0.2202 -0.0723  0.1717
#> 
#> Columns 19 to 20 -0.1376 -0.0939
#>  -0.1031  0.2631
#>   0.3316  0.4959
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4987  0.0816 -0.0114  0.4917 -0.3453 -0.0192  0.0884  0.0030 -0.2648
#>  -0.0747  0.0439 -0.4114  0.4348  0.4891 -0.0981  0.4463  0.0630  0.1613
#>  -0.3308  0.0891  0.0341 -0.0077 -0.4203  0.0047  0.2954 -0.1811 -0.5577
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2877 -0.6703  0.0122 -0.5237  0.7074  0.4558  0.0267  0.1334  0.0472
#>  -0.2040 -0.5775  0.3303 -0.2589  0.2794 -0.0216  0.2746  0.4775 -0.4032
#>   0.4093 -0.1342  0.2166 -0.5266  0.3200 -0.2445 -0.4722  0.3249 -0.1181
#> 
#> Columns 10 to 18 -0.0311 -0.5052 -0.1497 -0.2217  0.4092  0.4004  0.5404  0.0968 -0.4294
#>   0.0683 -0.4712 -0.5494 -0.4176  0.2616  0.2553  0.0171 -0.3827  0.2689
#>   0.5468 -0.5979 -0.0074  0.4572 -0.1937  0.1697 -0.0972 -0.0947 -0.5466
#> 
#> Columns 19 to 20  0.4713 -0.3400
#>   0.1951  0.3107
#>   0.6199 -0.4505
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6769 -0.4035  0.0207 -0.1323 -0.4497 -0.0861  0.2058  0.1945 -0.4685
#>  -0.3483 -0.0474 -0.2196  0.1040  0.0257  0.0797  0.2393  0.5542  0.2654
#>  -0.0769 -0.2145 -0.1048 -0.1349 -0.2803 -0.0365  0.0289  0.0396 -0.3003
#> 
#> Columns 10 to 18  0.6144 -0.2509  0.0562 -0.1991 -0.0288  0.0579 -0.6201  0.1129  0.4619
#>   0.2001 -0.3505  0.0928 -0.1906  0.0400  0.0380 -0.4467  0.0818  0.1468
#>   0.2341 -0.1839 -0.1803  0.2571  0.1573  0.0472 -0.3329  0.0340  0.4205
#> 
#> Columns 19 to 20  0.3224 -0.0475
#>   0.4150  0.3740
#>   0.2564  0.3749
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>