RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0933  0.3058  0.0189 -0.5585  0.1996  0.2716  0.1736  0.5452 -0.2235
#>   0.6890  0.4134  0.2741  0.0262  0.5953  0.5513  0.3596  0.0029  0.4361
#>   0.9731 -0.6739  0.8679  0.6144  0.9498 -0.7537 -0.3497  0.0181  0.0426
#> 
#> Columns 10 to 18 -0.7155 -0.5235 -0.6094 -0.6987 -0.1861 -0.1787  0.3181  0.3496 -0.4494
#>  -0.7181 -0.8329  0.0575  0.0154 -0.7075  0.6510  0.5927 -0.9418 -0.4339
#>  -0.2117  0.9020  0.2253 -0.6746  0.1574  0.1511  0.1277  0.2773 -0.1803
#> 
#> Columns 19 to 20  0.1727 -0.2713
#>  -0.5101 -0.4837
#>   0.6197  0.0577
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4617 -0.1323  0.2319  0.1536 -0.0604 -0.2461 -0.2997 -0.1960 -0.1504
#>   0.7212 -0.1385  0.0065 -0.2429 -0.0280 -0.4283 -0.6419  0.3508 -0.2185
#>   0.2792 -0.3860  0.1237  0.0128 -0.3980  0.2059 -0.0103  0.2162  0.0234
#> 
#> Columns 10 to 18 -0.5502 -0.2630 -0.6130 -0.2892 -0.6021 -0.2921  0.5751 -0.0493  0.2275
#>   0.0454 -0.0520 -0.6425  0.1858  0.3280  0.1414 -0.0683 -0.1613 -0.2505
#>  -0.4985 -0.2181 -0.5049 -0.3011 -0.1775 -0.5953  0.6263 -0.2581  0.2199
#> 
#> Columns 19 to 20  0.4625  0.1921
#>   0.2014  0.0472
#>   0.6780 -0.3585
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3362 -0.5982  0.3320 -0.1266 -0.2442  0.1968 -0.2577 -0.1583  0.0775
#>   0.3129 -0.3932 -0.0124 -0.2731  0.2161 -0.5481  0.1023  0.3830  0.2555
#>  -0.1033 -0.6031 -0.3815 -0.6623 -0.0232 -0.3147 -0.1561 -0.1182  0.2751
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5444  0.5985  0.3206 -0.0256  0.6136  0.3903  0.5589 -0.1175  0.5306
#>  -0.5814 -0.1401  0.0178  0.7652  0.0564  0.5157  0.3863  0.0564  0.6624
#>  -0.3154  0.2125  0.3128  0.3166  0.4999  0.2595 -0.0247 -0.4008  0.5501
#> 
#> Columns 10 to 18 -0.7753  0.0589 -0.1391  0.2926  0.4688  0.0615 -0.3549 -0.1473  0.3292
#>  -0.2180  0.3023 -0.2054  0.5696  0.5328 -0.0403 -0.3759  0.0611 -0.4867
#>  -0.5761  0.4811  0.4251  0.0976  0.6215  0.0654 -0.4626  0.3116  0.0409
#> 
#> Columns 19 to 20 -0.1270 -0.7284
#>  -0.0582 -0.4705
#>  -0.0427 -0.0118
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0129 -0.4711  0.0793 -0.3724 -0.1773 -0.1108  0.0253 -0.3203  0.0167
#>  -0.4528 -0.4301 -0.4527 -0.6275  0.0791 -0.2828  0.2613  0.0850  0.0190
#>   0.0102 -0.3599 -0.0678 -0.5412  0.0225 -0.3301 -0.1051  0.2176  0.2387
#> 
#> Columns 10 to 18  0.2278 -0.0295 -0.6620 -0.2580 -0.2550 -0.5717  0.2834 -0.6248  0.0670
#>   0.2818 -0.1277 -0.3656 -0.0355 -0.2297 -0.0634 -0.0277 -0.5853 -0.0673
#>   0.2938 -0.2877 -0.4784 -0.1014 -0.2895  0.1026  0.4374 -0.4504  0.0903
#> 
#> Columns 19 to 20  0.1598  0.1438
#>   0.1295  0.2887
#>   0.3219  0.3278
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>