RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0711 -0.0073  0.0878 -0.7275  0.6217  0.1210  0.6246 -0.6542  0.0167
#>  -0.4110  0.0139 -0.6814  0.1801  0.4952  0.7856  0.7391  0.2485 -0.1019
#>   0.4624  0.1428  0.4745 -0.4243  0.0897 -0.5752 -0.8245  0.0058 -0.4705
#> 
#> Columns 10 to 18  0.0516  0.8433  0.6888  0.2600  0.2254 -0.4816 -0.7199  0.4041 -0.5160
#>  -0.4053  0.8348  0.1791 -0.2991  0.6450  0.7301  0.8175 -0.5271  0.5811
#>  -0.1458  0.6397 -0.6144 -0.2480  0.4905  0.8687 -0.6670 -0.0894 -0.7134
#> 
#> Columns 19 to 20  0.8857  0.5462
#>   0.0054 -0.0487
#>   0.2130 -0.7472
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2958  0.2632 -0.1581  0.0754  0.3873 -0.0389 -0.0360 -0.0903 -0.0475
#>  -0.0475  0.0472 -0.5358  0.1843 -0.0352 -0.0912  0.2202  0.5059  0.3292
#>   0.2467 -0.1968  0.2352  0.4229  0.2503  0.0140 -0.3997  0.4281  0.3378
#> 
#> Columns 10 to 18  0.1495 -0.3183 -0.5475 -0.6527 -0.2329  0.1909  0.3867 -0.1918  0.3783
#>   0.0941 -0.6579 -0.3115 -0.6484 -0.5147 -0.0617  0.7147 -0.0448  0.4051
#>  -0.1630  0.5593  0.2584  0.6212 -0.1690  0.4568  0.1096 -0.1681 -0.4615
#> 
#> Columns 19 to 20  0.4323 -0.2170
#>   0.1183  0.0686
#>  -0.4362  0.0322
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2741 -0.4022 -0.2366  0.3287 -0.0637  0.0510  0.0748  0.0055  0.1987
#>   0.3509 -0.3861 -0.4935 -0.3110 -0.2817  0.1137  0.4766 -0.3214 -0.0983
#>  -0.3523  0.3968 -0.1676 -0.0604  0.3757 -0.0232  0.0910  0.5209  0.0761
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1104  0.0588  0.2697  0.1639  0.3192 -0.2387  0.2055  0.1661  0.3807
#>   0.1736 -0.5118 -0.5798 -0.7768 -0.3850 -0.0341  0.1296  0.0178  0.4335
#>  -0.2377  0.2439  0.1379  0.8517  0.7180 -0.2511 -0.8981  0.0146  0.0282
#> 
#> Columns 10 to 18  0.4032  0.3037 -0.1500 -0.7668  0.3086  0.2494 -0.4271 -0.1088  0.4045
#>  -0.2824 -0.5010 -0.5286 -0.0166  0.1517  0.5930 -0.2978  0.3406 -0.6445
#>   0.4601 -0.1769 -0.0579 -0.8968  0.5661  0.5773 -0.3647 -0.3933 -0.6610
#> 
#> Columns 19 to 20 -0.4616  0.2780
#>  -0.0070  0.1049
#>  -0.1122  0.0157
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2942 -0.2009 -0.2644  0.1799 -0.2161 -0.0235 -0.4256  0.1402  0.1067
#>   0.2048 -0.1492  0.4461  0.1086  0.2250  0.2829  0.2208  0.2275 -0.1657
#>   0.1954 -0.2576 -0.4715 -0.3338 -0.2186  0.0783 -0.2333  0.0566  0.2177
#> 
#> Columns 10 to 18 -0.0654  0.0895 -0.0217 -0.1248  0.0858  0.3070 -0.4082  0.2364 -0.0797
#>  -0.1352 -0.0597 -0.0682  0.1703  0.3887  0.1332  0.2400  0.1352  0.1115
#>   0.0316  0.1064 -0.3431  0.3093  0.1564  0.1400  0.1261  0.2199  0.0003
#> 
#> Columns 19 to 20  0.3344  0.1865
#>  -0.1128  0.1789
#>   0.4411  0.4488
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>