RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8239 -0.9844 -0.5758  0.9496 -0.2275  0.4905  0.2097  0.3948 -0.2153
#>   0.1532 -0.7916  0.5574 -0.2820  0.4557  0.2781  0.2553  0.7734  0.5794
#>  -0.7814 -0.8197 -0.3269  0.9040  0.8607 -0.5542  0.3530 -0.2592 -0.0372
#> 
#> Columns 10 to 18 -0.0354 -0.4116 -0.7429  0.2429  0.3023  0.5074 -0.3882  0.9440  0.0244
#>   0.3714  0.4317  0.6300  0.2022 -0.1744 -0.8073  0.5027  0.4546  0.6298
#>   0.0111  0.6155 -0.8770 -0.5514  0.6758 -0.6608 -0.9176  0.9245 -0.7233
#> 
#> Columns 19 to 20 -0.6674  0.2625
#>  -0.1784 -0.5878
#>  -0.6502 -0.9582
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1922 -0.3232  0.0680  0.0753  0.6723 -0.1569  0.2004  0.5811  0.1727
#>  -0.1588 -0.3967  0.0035  0.5706  0.0564 -0.4752 -0.5818  0.0495 -0.3636
#>   0.5806  0.1080 -0.7248  0.3326  0.4049 -0.3676  0.3679 -0.5594 -0.0023
#> 
#> Columns 10 to 18  0.7718  0.8685  0.3829  0.2019 -0.3598 -0.4344  0.3250  0.1750  0.4985
#>   0.1947 -0.1584 -0.1783 -0.6645 -0.5325  0.1331 -0.5452  0.6150  0.3263
#>   0.8346  0.2463 -0.2045 -0.3589 -0.5830 -0.4626  0.0682  0.0734 -0.2282
#> 
#> Columns 19 to 20 -0.3729  0.0267
#>   0.3265  0.5564
#>  -0.6563  0.4119
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5516 -0.4506  0.3648  0.4221  0.1357 -0.2882  0.2069 -0.2733  0.2019
#>  -0.0695 -0.3571  0.2308  0.1382  0.5953 -0.6733  0.2494 -0.3899  0.0327
#>   0.1502 -0.4149 -0.3584  0.1668  0.6679 -0.4648 -0.1479 -0.0205 -0.0501
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1098 -0.3757  0.2462 -0.4867 -0.5267  0.4353 -0.0002 -0.2150 -0.2823
#>  -0.1430  0.5404 -0.5567 -0.0505 -0.7811 -0.4542  0.3986 -0.0915 -0.4518
#>   0.2554  0.7780 -0.2466  0.2739 -0.3944 -0.2221 -0.1412  0.1152  0.0335
#> 
#> Columns 10 to 18  0.1999  0.6496  0.3430  0.7056 -0.3683  0.1313 -0.2142  0.3194 -0.4731
#>   0.5910  0.5826  0.2129  0.6011 -0.3611 -0.0680 -0.3982  0.0689 -0.1725
#>   0.1677  0.5984 -0.0374  0.0146 -0.3263 -0.6481 -0.0977 -0.1124  0.1808
#> 
#> Columns 19 to 20  0.2105 -0.7052
#>  -0.1971 -0.1953
#>   0.3803 -0.2325
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1285 -0.2933 -0.1074  0.2833  0.5476 -0.4333  0.3055  0.0476 -0.0023
#>   0.0210 -0.0528 -0.2860 -0.1778  0.4475 -0.6549  0.0152  0.0347 -0.1080
#>   0.1357 -0.1705 -0.4332 -0.2462  0.6091 -0.6619 -0.0790 -0.3320 -0.1981
#> 
#> Columns 10 to 18  0.1604  0.6248 -0.0070 -0.4386  0.4235 -0.7308  0.0113  0.0491  0.1536
#>  -0.0495  0.7615  0.4677 -0.4826 -0.2764 -0.4374  0.1234  0.1096  0.5099
#>   0.4300  0.5195  0.1898 -0.5392 -0.0692 -0.4474  0.0786  0.6478 -0.0232
#> 
#> Columns 19 to 20  0.1784 -0.0753
#>  -0.1500  0.3290
#>  -0.2829  0.1934
#> [ CPUFloatType{2,3,20} ]
#>