RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.2748 -0.7731  0.6574 -0.1834  0.2649  0.2147  0.6192  0.5657 -0.7684
#>  -0.7135 -0.3000 -0.2514 -0.4300 -0.7108  0.4822 -0.2862 -0.5967 -0.0011
#>  -0.5032  0.7446 -0.0801  0.5606  0.3016 -0.2728  0.4575  0.1746  0.9494
#> 
#> Columns 10 to 18  0.7971  0.8920  0.3606 -0.8779  0.4088  0.9326  0.3361  0.4627  0.0739
#>  -0.3713 -0.4414 -0.0177  0.0054  0.1034  0.0176 -0.7223  0.5243  0.9365
#>  -0.0037  0.0130  0.1746 -0.2561  0.1075 -0.5920 -0.3309  0.0249  0.3343
#> 
#> Columns 19 to 20 -0.6977 -0.6235
#>   0.8030  0.0877
#>   0.3079  0.2873
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2283  0.3537  0.3588  0.8187 -0.6345  0.2247  0.0181  0.5032 -0.6098
#>  -0.0909 -0.4598  0.4005 -0.4627 -0.1436  0.1859 -0.7259 -0.1608 -0.1182
#>  -0.6256  0.2215  0.0243  0.4329 -0.5385  0.1861 -0.0567 -0.1901  0.2271
#> 
#> Columns 10 to 18  0.7175 -0.7899 -0.2084 -0.4382 -0.0943 -0.0348  0.1372  0.6242  0.7746
#>   0.4809 -0.0759 -0.0901  0.2210 -0.3929 -0.0867 -0.3507  0.2131  0.2586
#>   0.4444 -0.2399 -0.3022 -0.2210  0.6826 -0.5177 -0.4905  0.1248 -0.0691
#> 
#> Columns 19 to 20 -0.2881 -0.1245
#>   0.0454  0.3662
#>  -0.2678  0.2414
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0780  0.3097  0.3786  0.4409 -0.7129 -0.3054 -0.6793 -0.1371 -0.0850
#>  -0.1854 -0.0671  0.1503  0.2401  0.0528 -0.0403 -0.4081  0.2057 -0.6071
#>   0.3380  0.2626  0.5608  0.0323 -0.4657 -0.3236 -0.2804  0.1915 -0.0409
#> 
#> Columns 10 to 18  0.6387 -0.1154 -0.2153  0.1790  0.0989 -0.0423 -0.4810  0.0821  0.2284
#>   0.3755  0.0471 -0.1040 -0.5225 -0.0471  0.3278  0.1619 -0.0656  0.3607
#>   0.5354  0.2545  0.1839 -0.2445  0.1415  0.1813 -0.2869 -0.0803  0.6671
#> 
#> Columns 19 to 20 -0.0984 -0.3141
#>  -0.0735  0.1540
#>   0.3892 -0.1642
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.1751  0.1919  0.3069  0.1600 -0.3407 -0.5520 -0.4760  0.0634 -0.3102
#>  -0.3711  0.3208  0.2381  0.3995 -0.5315  0.2040 -0.3124  0.0070 -0.2937
#>  -0.4758 -0.0540  0.3812  0.3727 -0.2016 -0.1833 -0.0643  0.2266 -0.1752
#> 
#> Columns 10 to 18  0.8582  0.5124 -0.2293 -0.1705  0.2377  0.6146 -0.3876  0.3000  0.2641
#>   0.1828 -0.4294 -0.2086 -0.3415 -0.1535 -0.1306  0.2832  0.1272  0.3884
#>   0.5763 -0.3058 -0.6874 -0.4998 -0.3374  0.1235 -0.0824  0.3454  0.1899
#> 
#> Columns 19 to 20  0.4537  0.1577
#>  -0.5067 -0.1196
#>  -0.0605  0.3060
#> 
#> (5,.,.) = 
#>  Columns 1 to 9 -0.7028 -0.0932  0.1769  0.6676 -0.3329 -0.1384 -0.4127 -0.1537 -0.4679
#>   0.0917  0.0786  0.3402  0.3392 -0.3226 -0.7059 -0.4524  0.2378  0.0020
#>  -0.2221  0.4678  0.3108  0.4164 -0.2792 -0.2515 -0.6533  0.0898  0.0342
#> 
#> Columns 10 to 18  0.4406 -0.4131 -0.4801 -0.2822  0.3398  0.3357 -0.1450  0.1423  0.7369
#>   0.7501  0.2470 -0.1773 -0.1880 -0.1691  0.2330 -0.5197 -0.3162  0.5937
#>   0.6107 -0.1459 -0.2188 -0.1389  0.0344  0.0143 -0.1021 -0.0764  0.6095
#> 
#> Columns 19 to 20 -0.1985  0.1239
#>   0.0153 -0.0557
#>  -0.2562 -0.2798
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5563 -0.0686 -0.7849 -0.1578 -0.2666 -0.5782 -0.8468  0.1376  0.1140
#>   0.1327 -0.7157 -0.5529 -0.2460 -0.3460  0.2402 -0.2178  0.2640 -0.7531
#>  -0.5467  0.1355 -0.3878  0.1960 -0.2926 -0.2143 -0.3186  0.4326 -0.5021
#> 
#> Columns 10 to 18 -0.7549 -0.2750 -0.0163  0.0815  0.3136  0.0594  0.0210  0.1176 -0.2401
#>  -0.2794 -0.2214 -0.5490 -0.0060  0.6274  0.3373 -0.7849  0.1967 -0.0034
#>  -0.8405 -0.2799 -0.0192  0.1770  0.7691 -0.3961  0.0834  0.0424  0.2579
#> 
#> Columns 19 to 20 -0.2681 -0.6596
#>  -0.0186 -0.1321
#>  -0.1685 -0.2538
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7028 -0.0932  0.1769  0.6676 -0.3329 -0.1384 -0.4127 -0.1537 -0.4679
#>   0.0917  0.0786  0.3402  0.3392 -0.3226 -0.7059 -0.4524  0.2378  0.0020
#>  -0.2221  0.4678  0.3108  0.4164 -0.2792 -0.2515 -0.6533  0.0898  0.0342
#> 
#> Columns 10 to 18  0.4406 -0.4131 -0.4801 -0.2822  0.3398  0.3357 -0.1450  0.1423  0.7369
#>   0.7501  0.2470 -0.1773 -0.1880 -0.1691  0.2330 -0.5197 -0.3162  0.5937
#>   0.6107 -0.1459 -0.2188 -0.1389  0.0344  0.0143 -0.1021 -0.0764  0.6095
#> 
#> Columns 19 to 20 -0.1985  0.1239
#>   0.0153 -0.0557
#>  -0.2562 -0.2798
#> [ CPUFloatType{2,3,20} ]
#>