RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4460 -0.5119 -0.4723  0.4357 -0.6968 -0.2844  0.4230 -0.1767 -0.0471
#>   0.0733 -0.8555 -0.9119 -0.0903 -0.4503  0.6005 -0.7814 -0.2288 -0.6043
#>  -0.7508 -0.6117 -0.5958 -0.7703  0.1257 -0.7715 -0.5173  0.1231 -0.6298
#> 
#> Columns 10 to 18  0.5387 -0.1723 -0.5172 -0.3691  0.5637  0.2061 -0.3819 -0.0419 -0.1251
#>   0.3191  0.9003 -0.4163  0.4466 -0.2094  0.3524 -0.1973 -0.2072  0.3488
#>   0.8439 -0.0476  0.5835  0.5592  0.1673  0.2653 -0.7777 -0.2223  0.4383
#> 
#> Columns 19 to 20 -0.5568 -0.2843
#>   0.2522  0.4910
#>   0.1755  0.4800
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3364 -0.3527 -0.5007 -0.3554 -0.3117  0.6074 -0.3063  0.4194  0.2002
#>   0.1676  0.3369 -0.8042 -0.3208 -0.1164  0.3610  0.0465  0.6633  0.0694
#>   0.1688  0.3596 -0.6575 -0.0398 -0.1120  0.3263  0.0785  0.4124 -0.2307
#> 
#> Columns 10 to 18  0.4446  0.0436 -0.3373  0.3509  0.0274  0.0579  0.1819  0.1083 -0.6493
#>   0.4895  0.5927  0.1861  0.6421  0.2649  0.0699 -0.1700 -0.0183  0.2001
#>   0.5842  0.0956 -0.0100  0.4847  0.0238  0.7752 -0.3305  0.1499 -0.0007
#> 
#> Columns 19 to 20  0.3294 -0.3986
#>   0.4546  0.1136
#>  -0.1201  0.3547
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0494  0.0955  0.0227 -0.2012 -0.1501  0.1968 -0.1834  0.0145  0.0892
#>   0.0470  0.5049  0.0749 -0.1359 -0.3432  0.5802  0.1290  0.0960  0.0697
#>   0.2347 -0.4195  0.2038 -0.6416  0.1158  0.1872 -0.6092  0.2247  0.0593
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0198 -0.6328 -0.5086 -0.0852 -0.2263 -0.3697  0.1830  0.4928  0.3642
#>  -0.1080 -0.8784  0.4589 -0.2559  0.5547 -0.0893 -0.1974  0.4497  0.2674
#>  -0.2933 -0.8660 -0.1593  0.4703  0.3155 -0.4733  0.4243 -0.2406 -0.3942
#> 
#> Columns 10 to 18 -0.1832 -0.3048  0.4781  0.5507  0.1214  0.1532 -0.1700 -0.1188 -0.4919
#>   0.3037  0.3713  0.1458  0.3594  0.6752 -0.5135 -0.5712 -0.1623 -0.0942
#>  -0.0423 -0.0601  0.0087  0.4562 -0.3009 -0.3011  0.0698 -0.1764 -0.2972
#> 
#> Columns 19 to 20 -0.5134  0.3971
#>  -0.5028  0.3120
#>   0.2347 -0.0098
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3507 -0.0856 -0.1608 -0.2332 -0.0400  0.3568  0.2024  0.1518  0.4793
#>  -0.1953 -0.1374 -0.4402 -0.2332  0.2802 -0.0544 -0.1590  0.2223 -0.0603
#>   0.2970 -0.2188 -0.3809  0.0114 -0.2178  0.2684 -0.2808  0.2886  0.1337
#> 
#> Columns 10 to 18  0.0563  0.0626 -0.0627  0.3443 -0.0681 -0.1726 -0.3228 -0.0055 -0.5203
#>   0.4786  0.2888  0.2730  0.3281  0.0017  0.0739 -0.3670 -0.0975  0.0754
#>   0.7159  0.3739 -0.0023  0.6212 -0.0119  0.3954 -0.0703  0.3357 -0.0931
#> 
#> Columns 19 to 20  0.0993 -0.1789
#>  -0.0476  0.2475
#>  -0.3851  0.2848
#> [ CPUFloatType{2,3,20} ]
#>