RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\text{ReLU}\) non-linearity to an input sequence.

nn_rnn(...)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional:
If TRUE, becomes a bidirectional RNN. Default: FALSE

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\text{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\text{input_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\text{hidden_size}\) Defaults to zero if not provided. where \(S=\text{num_layers} * \text{num_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\text{num_directions} * \text{hidden_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_lk: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_lk: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_lk: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_lk: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\text{hidden\_size}}\)

Examples

rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.9483  0.7461  0.7179  0.5935 -0.1447 -0.4223  0.3253  0.2011  0.6209
#>   0.0838 -0.0133 -0.3703 -0.7828  0.7113  0.8314 -0.3609  0.2330 -0.2860
#>  -0.3281 -0.7516 -0.5350  0.8472 -0.8357  0.5996 -0.4021  0.6833  0.8228
#> 
#> Columns 10 to 18 -0.1935  0.2527  0.3583 -0.0219  0.4799  0.1662  0.1633 -0.0902  0.7871
#>   0.4915  0.5770 -0.0657 -0.6724 -0.4763 -0.4696 -0.4609 -0.2941  0.4487
#>  -0.5493  0.2063 -0.0472 -0.3378 -0.8615  0.9312  0.1280 -0.4593 -0.3969
#> 
#> Columns 19 to 20 -0.7195  0.8536
#>  -0.5420  0.4867
#>  -0.0975  0.1708
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5128 -0.0314  0.2695  0.0391 -0.2772  0.1860 -0.4123 -0.3513 -0.5344
#>   0.0941  0.5370  0.6698 -0.1626 -0.0433  0.1887 -0.3346  0.4612  0.0761
#>   0.3915 -0.3066  0.3751 -0.0348 -0.2602  0.2764  0.3399 -0.4852  0.0331
#> 
#> Columns 10 to 18  0.5372  0.3438  0.4199  0.0888 -0.4434  0.8941 -0.0878 -0.1663  0.3638
#>   0.5906  0.0826  0.7347 -0.5004 -0.3429  0.3652 -0.4580 -0.8003  0.2789
#>   0.2720  0.0467  0.4936  0.1007  0.0156 -0.3135 -0.3395  0.2615  0.2785
#> 
#> Columns 19 to 20  0.2648  0.2991
#>  -0.5519  0.2538
#>  -0.1506  0.3654
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5811 -0.4110 -0.0699 -0.0141 -0.2641  0.1260  0.0484  0.1292  0.3029
#>   0.2188 -0.5507 -0.3106 -0.0310 -0.3795  0.5876  0.3275 -0.1617  0.4083
#>   0.0141  0.3285  0.4161  0.0208  0.1986 -0.2962 -0.2526  0.1994  0.0280
#> 
#> Columns 10 to 18 -0.1692 -0.1383  0.0215 -0.1326 -0.0884 -0.0368 -0.2598  0.2711 -0.0129
#>  -0.4849  0.2877 -0.0820 -0.7216 -0.7220 -0.0496 -0.3973 -0.1436 -0.2840
#>  -0.1749 -0.2339  0.3988 -0.2391 -0.2474  0.4384 -0.2962 -0.4537  0.0771
#> 
#> Columns 19 to 20 -0.1319  0.0439
#>   0.1482 -0.5427
#>  -0.4426  0.2386
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.4323  0.2219  0.2488 -0.0950 -0.0378 -0.1112  0.1700 -0.1892  0.0510
#>   0.2067 -0.0118  0.5474  0.2571 -0.4433  0.1894  0.0620 -0.0214 -0.0977
#>   0.4223 -0.0692 -0.0091  0.0701 -0.3350 -0.0707  0.3578 -0.2199  0.5754
#> 
#> Columns 10 to 18 -0.2283  0.0101  0.2951 -0.0292 -0.0760  0.1667 -0.0236  0.0854  0.0128
#>   0.1826  0.3950  0.2878 -0.3538  0.2698  0.1488  0.3603 -0.0584 -0.1042
#>  -0.5628  0.2180 -0.0187 -0.3506 -0.6212  0.1005 -0.2351  0.0549 -0.1005
#> 
#> Columns 19 to 20 -0.3961  0.3290
#>  -0.3891  0.2817
#>  -0.1639  0.0967
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.2701  0.1818  0.2574 -0.0038  0.0514 -0.2156  0.0950 -0.1621  0.0350
#>   0.0399  0.3176  0.4758 -0.1502  0.0530  0.0735 -0.0663 -0.0668 -0.3787
#>  -0.1657  0.2797  0.4789  0.4622 -0.3996  0.1101 -0.2324 -0.0974  0.1868
#> 
#> Columns 10 to 18 -0.3353  0.0557  0.0231  0.1278 -0.3761  0.4854 -0.0519  0.1994 -0.1867
#>   0.2843  0.0144  0.6096 -0.3416  0.0373  0.4531 -0.1596 -0.7714  0.4894
#>  -0.0050  0.1386  0.1854 -0.4752 -0.2953  0.3228 -0.0686 -0.3481 -0.1328
#> 
#> Columns 19 to 20 -0.2346  0.5273
#>  -0.6407  0.3038
#>  -0.4549  0.0337
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.2017  0.7241 -0.2314  0.4396  0.0304 -0.2661 -0.3124  0.5995  0.8125
#>   0.3396  0.7052  0.6553 -0.6963  0.2436 -0.1440  0.5031 -0.3775  0.5127
#>   0.1128  0.1069  0.0075 -0.5158 -0.1782  0.2460  0.2991  0.2701 -0.2640
#> 
#> Columns 10 to 18 -0.2982 -0.6315 -0.1632  0.8407 -0.3253 -0.0332  0.3336 -0.2142 -0.8586
#>   0.3851  0.1568  0.4940  0.5739  0.4019 -0.2286  0.1499 -0.1344 -0.4895
#>  -0.0905 -0.4991  0.0363 -0.2359  0.1833 -0.2315  0.0707 -0.7023  0.1650
#> 
#> Columns 19 to 20 -0.5501  0.4365
#>   0.4423 -0.4196
#>   0.1555 -0.3474
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2701  0.1818  0.2574 -0.0038  0.0514 -0.2156  0.0950 -0.1621  0.0350
#>   0.0399  0.3176  0.4758 -0.1502  0.0530  0.0735 -0.0663 -0.0668 -0.3787
#>  -0.1657  0.2797  0.4789  0.4622 -0.3996  0.1101 -0.2324 -0.0974  0.1868
#> 
#> Columns 10 to 18 -0.3353  0.0557  0.0231  0.1278 -0.3761  0.4854 -0.0519  0.1994 -0.1867
#>   0.2843  0.0144  0.6096 -0.3416  0.0373  0.4531 -0.1596 -0.7714  0.4894
#>  -0.0050  0.1386  0.1854 -0.4752 -0.2953  0.3228 -0.0686 -0.3481 -0.1328
#> 
#> Columns 19 to 20 -0.2346  0.5273
#>  -0.6407  0.3038
#>  -0.4549  0.0337
#> [ CPUFloatType{2,3,20} ]
#>