RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3345 -0.8092  0.8616 -0.3814 -0.6915  0.4852  0.7867 -0.5962  0.9033
#>   0.2259  0.3547  0.1961  0.2019 -0.2665  0.6190  0.4105  0.9266 -0.1081
#>   0.6620 -0.2637  0.4504  0.3002 -0.1003 -0.9604 -0.4974  0.3616  0.2804
#> 
#> Columns 10 to 18 -0.2662  0.4342  0.0859  0.6060 -0.7298  0.4275 -0.5170 -0.6527  0.0447
#>  -0.1489 -0.1532 -0.3676  0.7770  0.3155 -0.4149  0.4342 -0.2582 -0.4781
#>  -0.5659 -0.4697 -0.7085 -0.8518  0.8653 -0.0720  0.1830 -0.4675 -0.8121
#> 
#> Columns 19 to 20  0.8197 -0.8390
#>   0.7136  0.8726
#>  -0.6550  0.9490
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1877 -0.0359  0.3635  0.1780  0.2971 -0.9101  0.4571 -0.4257  0.6578
#>  -0.0443 -0.2512  0.7624 -0.0894 -0.7454 -0.1553  0.0990  0.1471  0.1659
#>  -0.1011 -0.2845  0.2408  0.2113 -0.3587  0.3907  0.6455 -0.4355 -0.0128
#> 
#> Columns 10 to 18 -0.2258  0.1592  0.4052 -0.3029  0.6346 -0.3514 -0.3189 -0.3640 -0.4152
#>  -0.3894  0.3147 -0.3049  0.2987  0.7429 -0.3471  0.1917  0.2159  0.0276
#>  -0.1765  0.4714 -0.3878 -0.1281 -0.3013 -0.0446  0.2997  0.2569  0.1633
#> 
#> Columns 19 to 20  0.1087 -0.4652
#>  -0.0069  0.2470
#>   0.8382 -0.4049
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3968 -0.1327  0.4892  0.3467 -0.0534 -0.4809  0.6823 -0.4104  0.2749
#>   0.1936  0.4276  0.5926 -0.2124  0.0078 -0.5926  0.6900 -0.1461  0.2141
#>   0.0320  0.1210  0.3006  0.0531 -0.0199 -0.3555  0.2782 -0.2561  0.6810
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3201 -0.0094 -0.5913 -0.2042  0.2034 -0.2921 -0.2716  0.1504  0.0855
#>   0.0986  0.3801 -0.7451 -0.0820  0.5822 -0.5968 -0.1580 -0.0045  0.2787
#>   0.4021  0.3402 -0.5956  0.0071 -0.0387 -0.4006  0.1451 -0.3226  0.3752
#> 
#> Columns 10 to 18 -0.1617 -0.0465  0.0560 -0.5249 -0.2707  0.7919 -0.4247  0.2552  0.1553
#>  -0.3183  0.0216 -0.2405 -0.4657 -0.6597  0.7781 -0.7501 -0.0955  0.5029
#>   0.1261 -0.3039 -0.4118 -0.5502 -0.2039  0.4801 -0.3519  0.0634  0.0357
#> 
#> Columns 19 to 20  0.4430 -0.2941
#>  -0.5167 -0.5764
#>   0.5123 -0.5078
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2787 -0.1695  0.3868  0.0903 -0.3336 -0.6068  0.4856 -0.2817  0.2359
#>   0.2348 -0.1015  0.6116 -0.0421 -0.1798 -0.1931  0.4836 -0.4568  0.5772
#>   0.3405  0.1997  0.5214  0.1526 -0.0931 -0.1824  0.6392 -0.1871  0.3944
#> 
#> Columns 10 to 18 -0.3941 -0.1537 -0.2288 -0.2495  0.2253  0.0120 -0.2736  0.1356 -0.3282
#>  -0.2815  0.0964 -0.5073  0.0956 -0.1618 -0.0885 -0.1228 -0.0379 -0.2024
#>  -0.5154 -0.1136 -0.4561 -0.2810 -0.0525 -0.0471 -0.0752  0.0108 -0.1525
#> 
#> Columns 19 to 20  0.0616  0.0758
#>  -0.0229 -0.0540
#>  -0.0506 -0.0383
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>