RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6080 -0.4723  0.8124 -0.2303 -0.8406  0.4586 -0.5130 -0.7567 -0.5777
#>   0.0172  0.4680 -0.1801  0.6967 -0.7151  0.3258 -0.8904  0.2205 -0.5161
#>  -0.2895  0.4276  0.0639  0.5464  0.5609 -0.8782 -0.0109 -0.5712  0.8439
#> 
#> Columns 10 to 18  0.8438  0.9146  0.2830 -0.6760 -0.6261  0.5633  0.4315 -0.1131 -0.8019
#>   0.7037  0.8563 -0.6013  0.3537  0.9114  0.9414 -0.7021 -0.3818 -0.3197
#>   0.6900  0.5140 -0.1238 -0.7767 -0.0017  0.5883 -0.1074 -0.4426  0.3829
#> 
#> Columns 19 to 20  0.5673 -0.9236
#>  -0.2120 -0.5876
#>  -0.8873 -0.1068
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2314  0.3780 -0.5503 -0.3426  0.4899 -0.0464 -0.0851 -0.4534 -0.3796
#>   0.2252  0.6224  0.4087  0.1596 -0.2437  0.4395 -0.6274 -0.2968 -0.4088
#>  -0.2802  0.2490 -0.2967  0.3699 -0.3128  0.3638 -0.7438 -0.4413  0.1046
#> 
#> Columns 10 to 18  0.5928  0.6908 -0.3533 -0.4411  0.2072 -0.5655 -0.6964  0.2184 -0.1939
#>   0.7575  0.7953  0.5936 -0.2421 -0.3130  0.2535  0.1091  0.0434 -0.3287
#>   0.5486  0.7715  0.0700 -0.3124 -0.4636  0.2291  0.0211 -0.0994 -0.1305
#> 
#> Columns 19 to 20 -0.3664 -0.2418
#>   0.0637 -0.7982
#>   0.1960 -0.5038
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2291  0.0003  0.2598 -0.3257  0.3034 -0.3900 -0.1308 -0.0131  0.5363
#>  -0.0023  0.4795  0.3886  0.0164  0.5838 -0.0277 -0.2736 -0.6835  0.1891
#>   0.2398  0.4255 -0.2026 -0.3133  0.3932 -0.2853 -0.1438  0.1099  0.4108
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0548 -0.0785 -0.6603  0.7378  0.5130  0.0663 -0.4626  0.1797 -0.6119
#>   0.1206 -0.4977  0.2369  0.6320 -0.3122 -0.4885 -0.6673 -0.2853 -0.1050
#>   0.3375  0.3132 -0.9702  0.8637 -0.5014  0.5492  0.3042  0.4287 -0.6373
#> 
#> Columns 10 to 18 -0.7561  0.2938  0.0496 -0.6708 -0.9034  0.6936 -0.4530 -0.0648 -0.0934
#>   0.6331 -0.1800  0.1884 -0.6187  0.3845 -0.1808  0.5568  0.5574  0.2463
#>  -0.0383  0.2549  0.5401  0.4926 -0.0711 -0.6573 -0.3501 -0.2991  0.2362
#> 
#> Columns 19 to 20  0.3730  0.5472
#>   0.7229  0.4319
#>   0.6428 -0.3797
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5863  0.1147 -0.0200 -0.4075  0.5330 -0.3998 -0.0378  0.1014  0.6310
#>   0.2686  0.2969  0.3828 -0.0456  0.3917  0.1022 -0.6480 -0.3983  0.3600
#>  -0.7001  0.1080  0.5638  0.0913  0.3226 -0.0850 -0.5500 -0.3330  0.5163
#> 
#> Columns 10 to 18  0.0945  0.2880 -0.2684 -0.1644 -0.0592  0.2644 -0.3482  0.2527  0.4086
#>   0.4642  0.6529 -0.3559 -0.3873  0.2607 -0.3527 -0.1291 -0.2352 -0.2370
#>   0.3794  0.5844 -0.2354 -0.2038 -0.3263 -0.0301 -0.2056 -0.0216  0.0700
#> 
#> Columns 19 to 20 -0.0618  0.3633
#>  -0.0096  0.0034
#>  -0.4318 -0.2153
#> [ CPUFloatType{2,3,20} ]
#>