RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

# \dontrun{
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.1563  0.2797  0.4653  0.8483 -0.5044  0.5032 -0.2032  0.8317 -0.0350
#>   0.1632 -0.1904  0.5733  0.3198 -0.3562  0.3904 -0.6873  0.7901 -0.9472
#>  -0.6539 -0.6025  0.4797 -0.8375  0.7255 -0.4198  0.5030 -0.1419  0.2092
#> 
#> Columns 10 to 18  0.9481 -0.2581 -0.6185 -0.7787 -0.2685 -0.7144 -0.3700  0.4854  0.0008
#>   0.8805 -0.2755 -0.9071 -0.5580 -0.0061 -0.8662 -0.7541  0.7855 -0.6157
#>   0.5480  0.5395 -0.1759  0.2965  0.7805 -0.2651 -0.8080  0.2643  0.7689
#> 
#> Columns 19 to 20  0.3487 -0.1631
#>   0.2535  0.5113
#>  -0.0188  0.3766
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1416  0.4498  0.1932  0.1637 -0.5115 -0.7125 -0.4544 -0.1721 -0.0345
#>  -0.7202  0.1668  0.1129  0.6700 -0.6155  0.2947 -0.4558  0.1086  0.1323
#>   0.3006 -0.0124 -0.0014  0.5341 -0.0718  0.3542 -0.1502  0.4042 -0.5156
#> 
#> Columns 10 to 18  0.8057  0.2908 -0.1019 -0.2748  0.2799  0.6493 -0.7338  0.1910  0.1886
#>  -0.1154  0.1593 -0.5129 -0.6648  0.1733 -0.2496 -0.0886  0.2975  0.5137
#>   0.8810  0.4194 -0.0103 -0.8285 -0.0348 -0.0102 -0.6628  0.1669 -0.4613
#> 
#> Columns 19 to 20 -0.0403  0.0080
#>  -0.0600 -0.4162
#>   0.0908  0.6937
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1978  0.2242  0.0024  0.3932 -0.4801 -0.4895 -0.5400 -0.0527 -0.4520
#>   0.4544  0.0302  0.4917  0.2736 -0.5769 -0.1859 -0.4959  0.0229 -0.2535
#>  -0.5535  0.3675  0.5847  0.6636 -0.3288 -0.2481 -0.1065 -0.0289 -0.5147
#> 
#> Columns 10 to 18  0.8121  0.5476 -0.5889 -0.2491  0.5971  0.3482 -0.4202  0.5075  0.0695
#>   0.8887  0.3603 -0.1642 -0.3072  0.2559 -0.0096 -0.6545  0.5044  0.5036
#>   0.6373 -0.1207  0.0495 -0.3367  0.4293  0.4361 -0.3157  0.3224  0.6757
#> 
#> Columns 19 to 20 -0.0226 -0.0955
#>   0.6364  0.2054
#>   0.1772 -0.2871
#> 
#> (4,.,.) = 
#>  Columns 1 to 9 -0.5002  0.2480 -0.0165  0.4973 -0.7685  0.0885 -0.3330 -0.2697 -0.1477
#>  -0.5379  0.1719  0.2126  0.1891 -0.5105  0.2180 -0.5122 -0.1882 -0.4472
#>   0.0806  0.0901  0.5329  0.3643 -0.6769  0.4601 -0.5399 -0.3066 -0.0994
#> 
#> Columns 10 to 18  0.4677 -0.0194 -0.3609 -0.2897  0.3666  0.0276  0.0770  0.5985  0.5201
#>   0.6046 -0.2503 -0.4701 -0.1266  0.3423  0.1259 -0.2631  0.5912  0.1230
#>   0.6537 -0.2490 -0.3203 -0.3803  0.0304 -0.0077  0.1981  0.6495 -0.0583
#> 
#> Columns 19 to 20 -0.0329 -0.2124
#>   0.1306  0.0613
#>   0.0430 -0.0534
#> 
#> (5,.,.) = 
#>  Columns 1 to 9 -0.3560  0.0896  0.2468  0.0908 -0.3990 -0.1175 -0.3947 -0.0834  0.1421
#>   0.1891 -0.0772  0.2671  0.0296 -0.1929 -0.2009 -0.5507 -0.2240  0.2121
#>  -0.1833  0.2226 -0.0158  0.5592 -0.5925  0.0255 -0.6282 -0.1562  0.0561
#> 
#> Columns 10 to 18  0.5649 -0.0882 -0.4652 -0.2057  0.0088  0.0349 -0.0315  0.3252  0.6167
#>   0.5626  0.3505 -0.2768 -0.4894 -0.0599  0.4348 -0.1352  0.2022  0.2273
#>   0.2647  0.0037 -0.3756 -0.3976 -0.0172 -0.1532 -0.4150  0.3451  0.3110
#> 
#> Columns 19 to 20  0.2250  0.1283
#>  -0.0717  0.2627
#>   0.1909 -0.1445
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2977 -0.3901 -0.3494 -0.6523  0.3627  0.1448 -0.3341  0.2196  0.1126
#>   0.2888 -0.7529  0.1781 -0.0379 -0.2393  0.3807  0.1044 -0.0212 -0.5096
#>  -0.1402 -0.3835 -0.2036  0.3084  0.1285 -0.3805  0.1103  0.0476  0.2100
#> 
#> Columns 10 to 18 -0.6325 -0.1108 -0.1481  0.0602  0.7081 -0.3749  0.6918 -0.4901 -0.2858
#>   0.2888  0.4654  0.2154 -0.3173  0.4848  0.3496 -0.1522 -0.0645 -0.5102
#>   0.2073  0.5197  0.0807  0.4554  0.0247 -0.2980 -0.3274 -0.1698 -0.0551
#> 
#> Columns 19 to 20  0.0842  0.0867
#>   0.1297 -0.4188
#>   0.6599 -0.5773
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3560  0.0896  0.2468  0.0908 -0.3990 -0.1175 -0.3947 -0.0834  0.1421
#>   0.1891 -0.0772  0.2671  0.0296 -0.1929 -0.2009 -0.5507 -0.2240  0.2121
#>  -0.1833  0.2226 -0.0158  0.5592 -0.5925  0.0255 -0.6282 -0.1562  0.0561
#> 
#> Columns 10 to 18  0.5649 -0.0882 -0.4652 -0.2057  0.0088  0.0349 -0.0315  0.3252  0.6167
#>   0.5626  0.3505 -0.2768 -0.4894 -0.0599  0.4348 -0.1352  0.2022  0.2273
#>   0.2647  0.0037 -0.3756 -0.3976 -0.0172 -0.1532 -0.4150  0.3451  0.3110
#> 
#> Columns 19 to 20  0.2250  0.1283
#>  -0.0717  0.2627
#>   0.1909 -0.1445
#> [ CPUFloatType{2,3,20} ]
#> 

# }