RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0379 -0.0446 -0.5462  0.0735 -0.8058 -0.0669 -0.7949  0.4207  0.2634
#>  -0.4870  0.4060 -0.1216 -0.8255  0.7502 -0.7894 -0.0987  0.2994  0.2224
#>   0.2650  0.0258 -0.4853 -0.5605 -0.7954 -0.7066 -0.2508 -0.5983 -0.5166
#> 
#> Columns 10 to 18 -0.5750  0.0621 -0.5825  0.2005  0.0230  0.6000  0.4249 -0.9138  0.0693
#>   0.8222  0.5378  0.5406 -0.6812 -0.3322 -0.9178  0.1382 -0.0458 -0.5355
#>   0.1256  0.3592 -0.8461 -0.0635 -0.0189  0.0092  0.5049  0.1298  0.1893
#> 
#> Columns 19 to 20  0.2205  0.3593
#>  -0.4921 -0.8136
#>  -0.2981 -0.4979
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4150 -0.4482 -0.0955  0.4422 -0.0046 -0.1260 -0.7230 -0.1333 -0.0232
#>  -0.1470 -0.3297  0.3649  0.3382 -0.3929  0.2873  0.2657  0.4233  0.4573
#>  -0.6506  0.2498  0.0901  0.2718  0.1990 -0.7377 -0.5370 -0.5769 -0.0510
#> 
#> Columns 10 to 18 -0.3144  0.3390 -0.5619  0.1391 -0.1047  0.0437  0.3522 -0.2997  0.1449
#>  -0.0825  0.3193  0.1593  0.2290  0.0823  0.3451  0.0665 -0.5485  0.4886
#>  -0.4884 -0.3803 -0.1820  0.5255  0.1547 -0.4297  0.0956 -0.5246  0.2461
#> 
#> Columns 19 to 20 -0.1495 -0.3329
#>   0.0397 -0.2749
#>  -0.6421  0.3979
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4517  0.2904 -0.0908 -0.0310  0.3720 -0.6094 -0.2204 -0.2775 -0.0046
#>  -0.2038 -0.2604  0.5591  0.0495  0.1113  0.0763 -0.1556  0.1820  0.1538
#>   0.5101  0.2081 -0.2323  0.3119 -0.0934 -0.2180 -0.3058 -0.0095 -0.1008
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0780 -0.1379 -0.6290  0.4693  0.2181  0.2879 -0.4694 -0.5668  0.7880
#>  -0.2746  0.1811  0.4012  0.0351  0.0093  0.7893  0.1833 -0.1082 -0.0572
#>   0.0389  0.0082 -0.1979 -0.3469 -0.0243  0.2806  0.3705 -0.5535 -0.1885
#> 
#> Columns 10 to 18 -0.2022 -0.0729  0.0697  0.3106 -0.0575 -0.1294  0.3068 -0.3709 -0.5531
#>  -0.6605 -0.2401  0.5405 -0.0184 -0.4992  0.8344  0.1311 -0.5310 -0.2531
#>  -0.5258  0.0290  0.6442  0.3933 -0.2957  0.0453  0.3072 -0.3216 -0.0859
#> 
#> Columns 19 to 20 -0.1162 -0.5060
#>  -0.4812  0.2467
#>   0.3897 -0.0203
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1865 -0.2607  0.2399  0.0729 -0.3620 -0.1841 -0.0893  0.1794 -0.2036
#>  -0.3832  0.1766  0.7748  0.4774  0.0006  0.1283 -0.4681  0.7208  0.1033
#>   0.0962  0.0263 -0.0189 -0.0515  0.1046 -0.5950 -0.3968  0.2098 -0.1120
#> 
#> Columns 10 to 18 -0.0922  0.2435 -0.0051  0.0131  0.1129 -0.3379  0.1234 -0.2605  0.1730
#>  -0.2477  0.1991 -0.0325  0.0830 -0.4994 -0.1011  0.5103 -0.5591  0.0383
#>   0.0633  0.0561 -0.4613 -0.2496 -0.3163 -0.5122  0.2369 -0.0049  0.1006
#> 
#> Columns 19 to 20  0.0423 -0.1972
#>  -0.4562 -0.5485
#>  -0.2273 -0.3321
#> [ CPUFloatType{2,3,20} ]
#>