RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\text{ReLU}\) non-linearity to an input sequence.

nn_rnn(...)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional:
If TRUE, becomes a bidirectional RNN. Default: FALSE

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\text{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\text{input_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\text{hidden_size}\) Defaults to zero if not provided. where \(S=\text{num_layers} * \text{num_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\text{num_directions} * \text{hidden_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_lk: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_lk: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_lk: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_lk: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\text{hidden\_size}}\)

Examples

rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0450  0.2275 -0.4077  0.7221 -0.6667  0.1383 -0.7369  0.2237 -0.1367
#>   0.3671 -0.6049 -0.5849  0.6251 -0.2441  0.2619  0.0253 -0.0825 -0.0936
#>  -0.4253 -0.0698 -0.0761  0.2025 -0.5423  0.2784  0.3202  0.4417 -0.7740
#> 
#> Columns 10 to 18 -0.1700  0.0034 -0.8964 -0.9198 -0.7022 -0.3096  0.6773  0.2286 -0.8212
#>  -0.3435 -0.4357 -0.3306  0.1745  0.1221  0.3041  0.5086  0.4532 -0.0845
#>   0.2640  0.8148 -0.3550 -0.5655 -0.8097  0.1487 -0.4068 -0.4634  0.0614
#> 
#> Columns 19 to 20 -0.3440  0.5179
#>   0.7779  0.8953
#>   0.4660  0.6423
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3582 -0.7956  0.4462 -0.7726  0.0521  0.2200 -0.1152 -0.5630  0.3734
#>  -0.4799 -0.4202 -0.2155 -0.0588 -0.4309  0.4847 -0.3443 -0.5467 -0.0543
#>  -0.7544 -0.4532 -0.1407 -0.4278 -0.0827  0.2435 -0.2162 -0.3907 -0.1011
#> 
#> Columns 10 to 18  0.0736  0.3639 -0.2185  0.3749 -0.0111  0.7981 -0.2350 -0.0253  0.1595
#>   0.4333  0.6557 -0.4204 -0.3173  0.3908  0.5348  0.6021  0.2237 -0.1009
#>   0.4562 -0.0913 -0.3012 -0.1139  0.4183  0.3039 -0.2912 -0.3884  0.2192
#> 
#> Columns 19 to 20  0.4460 -0.5698
#>   0.4400 -0.0834
#>   0.6187 -0.1766
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3402 -0.2592 -0.5211  0.2471 -0.2847 -0.2706 -0.0609 -0.2234  0.2405
#>  -0.2662  0.0575 -0.4005 -0.1628  0.0863  0.2703 -0.1572  0.2288  0.6992
#>  -0.1689 -0.3406 -0.1957 -0.0160 -0.1255 -0.2871  0.0128 -0.2830  0.4683
#> 
#> Columns 10 to 18 -0.0034 -0.0486  0.2790 -0.2818 -0.2299 -0.0231  0.3832  0.0291  0.2660
#>   0.0408  0.2657  0.1230  0.1941  0.2390  0.0019  0.2926  0.2653 -0.1557
#>  -0.1236 -0.1120  0.0060 -0.1511 -0.1384  0.2748  0.4960 -0.0688  0.0340
#> 
#> Columns 19 to 20 -0.3132 -0.3465
#>  -0.2362  0.2123
#>  -0.0361  0.1588
#> 
#> (4,.,.) = 
#>  Columns 1 to 9 -0.2134 -0.7002 -0.0058 -0.2570 -0.3701  0.3556 -0.2925  0.0129  0.2546
#>  -0.0360 -0.4036  0.4746 -0.4007  0.1530 -0.3018  0.1495  0.1202  0.1550
#>  -0.3565 -0.7167  0.1709 -0.4197 -0.2487  0.1707 -0.1795 -0.0590 -0.0206
#> 
#> Columns 10 to 18  0.1315 -0.2278 -0.6795  0.3050  0.0342  0.4811 -0.0010 -0.0135 -0.2659
#>   0.0595  0.0181 -0.1329  0.6455 -0.2651  0.3084 -0.2344  0.0875  0.1125
#>   0.2097  0.0047 -0.2817  0.3009 -0.1316  0.3746 -0.0177  0.1325 -0.0471
#> 
#> Columns 19 to 20  0.4823  0.4343
#>   0.0368  0.1256
#>   0.3974 -0.1822
#> 
#> (5,.,.) = 
#>  Columns 1 to 9 -0.1355 -0.4083 -0.2001  0.0834 -0.4540  0.2423  0.0177 -0.6781  0.2902
#>  -0.4247 -0.0863  0.0213  0.3566 -0.4386  0.2214 -0.3811 -0.4175  0.3920
#>  -0.4470 -0.3697 -0.1741  0.0373 -0.1536 -0.1044 -0.0745 -0.2894  0.2854
#> 
#> Columns 10 to 18  0.0430  0.1788 -0.6262 -0.4004  0.0606  0.7225  0.3877 -0.0528  0.2496
#>   0.3322  0.0008 -0.1720  0.0401  0.3725  0.2212 -0.0147 -0.2648 -0.0608
#>   0.1549 -0.2045  0.2249 -0.1287 -0.0823  0.2605  0.0401  0.0610 -0.0826
#> 
#> Columns 19 to 20  0.2690  0.4394
#>   0.1857  0.1339
#>  -0.0976 -0.3507
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1757  0.4477  0.5398  0.3722  0.5698  0.6913  0.3120  0.8522 -0.4493
#>  -0.0649 -0.3889  0.4403  0.6093 -0.2854 -0.7365  0.0727 -0.4360  0.2342
#>  -0.3436 -0.0606 -0.1078  0.1138 -0.4242 -0.0799  0.4655 -0.0752  0.1212
#> 
#> Columns 10 to 18  0.3280  0.4952  0.2430  0.6829  0.2111  0.5742 -0.8084 -0.5070  0.4987
#>  -0.3675 -0.2563  0.6384  0.1112 -0.4150 -0.0470  0.3778 -0.0940 -0.2094
#>   0.0078 -0.3756  0.0949 -0.2207 -0.0863 -0.2936 -0.0224 -0.0871 -0.2699
#> 
#> Columns 19 to 20  0.7389  0.2508
#>  -0.0740  0.4142
#>  -0.2614  0.5583
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1355 -0.4083 -0.2001  0.0834 -0.4540  0.2423  0.0177 -0.6781  0.2902
#>  -0.4247 -0.0863  0.0213  0.3566 -0.4386  0.2214 -0.3811 -0.4175  0.3920
#>  -0.4470 -0.3697 -0.1741  0.0373 -0.1536 -0.1044 -0.0745 -0.2894  0.2854
#> 
#> Columns 10 to 18  0.0430  0.1788 -0.6262 -0.4004  0.0606  0.7225  0.3877 -0.0528  0.2496
#>   0.3322  0.0008 -0.1720  0.0401  0.3725  0.2212 -0.0147 -0.2648 -0.0608
#>   0.1549 -0.2045  0.2249 -0.1287 -0.0823  0.2605  0.0401  0.0610 -0.0826
#> 
#> Columns 19 to 20  0.2690  0.4394
#>   0.1857  0.1339
#>  -0.0976 -0.3507
#> [ CPUFloatType{2,3,20} ]
#>