RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2513  0.1182  0.5844  0.8495 -0.7948 -0.4959 -0.1505  0.0853  0.0840
#>  -0.0362  0.0118  0.4049 -0.4106  0.4406  0.8264  0.4929 -0.4938 -0.1065
#>  -0.4279 -0.5451 -0.9043  0.1429 -0.7661 -0.6102  0.4254  0.1829 -0.0343
#> 
#> Columns 10 to 18 -0.3932 -0.8851 -0.2867  0.1358 -0.7181 -0.6259 -0.0012 -0.9061 -0.6497
#>   0.4781 -0.1949  0.5001  0.3497 -0.1681  0.5162  0.6331  0.5258 -0.0759
#>   0.3154 -0.2234  0.3130 -0.5894  0.1851 -0.0820  0.8941  0.4558 -0.9320
#> 
#> Columns 19 to 20  0.8460 -0.2051
#>  -0.3982  0.6386
#>   0.4927 -0.5231
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2449  0.2675  0.1057  0.7797  0.0686 -0.5664  0.6155 -0.2611 -0.4780
#>  -0.0342  0.0890  0.6427  0.2321  0.4404  0.6151  0.7302 -0.1684 -0.4287
#>  -0.2642  0.2643 -0.1665  0.7427  0.2103 -0.3294  0.2296 -0.2809 -0.5573
#> 
#> Columns 10 to 18  0.5052 -0.0031 -0.2301 -0.6247 -0.1262 -0.2690  0.2816 -0.0079 -0.2788
#>   0.0295 -0.0569  0.5321  0.0161  0.2086  0.4631  0.1043  0.0061  0.5476
#>   0.1968  0.0004  0.6438  0.1119  0.0694 -0.4597  0.2270 -0.8261 -0.4350
#> 
#> Columns 19 to 20  0.4131  0.1386
#>   0.2457  0.4799
#>   0.0155  0.3527
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2328 -0.4886 -0.0069  0.5817 -0.1535 -0.2618 -0.0839  0.1219 -0.1492
#>   0.0613 -0.2866  0.0918  0.3880  0.4224  0.1976  0.3609  0.0613 -0.1224
#>  -0.1950 -0.2049  0.0461 -0.0621  0.1020 -0.1563  0.4237 -0.3719 -0.6714
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4265 -0.0398 -0.4125  0.6554 -0.6746 -0.6066 -0.1937  0.3415  0.4727
#>   0.1381 -0.5299  0.5680 -0.7252  0.6378  0.5266  0.3053  0.6146 -0.0952
#>  -0.0126 -0.2972  0.1406 -0.0512 -0.8190 -0.4772  0.0234  0.0620  0.6267
#> 
#> Columns 10 to 18  0.5269  0.1475  0.2675  0.2703 -0.5154  0.2105  0.3664 -0.0060 -0.4052
#>  -0.7053 -0.5717 -0.5580 -0.7434 -0.3648 -0.5661 -0.6362 -0.3749 -0.1107
#>  -0.4199  0.1893 -0.2353 -0.4411 -0.7484  0.1911  0.2742 -0.2714  0.3674
#> 
#> Columns 19 to 20  0.5120  0.1680
#>   0.1631  0.9094
#>  -0.5446  0.7966
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3003 -0.2802  0.1964  0.6283 -0.3858 -0.5974  0.0149  0.0194 -0.2654
#>   0.1106 -0.1318  0.0275 -0.3098  0.6933  0.5461  0.6065 -0.1149 -0.3828
#>   0.1175 -0.4074  0.2188  0.5916  0.3321 -0.1164  0.4845  0.0517 -0.0845
#> 
#> Columns 10 to 18  0.0936  0.2070  0.1921  0.0146  0.1422  0.2572  0.3870 -0.3517 -0.1977
#>   0.3934  0.4017  0.6801 -0.0683 -0.1450 -0.0430  0.0375  0.0331  0.0498
#>   0.0101  0.3278  0.2476 -0.2297  0.3160  0.2372  0.3320  0.4389 -0.1142
#> 
#> Columns 19 to 20  0.4774  0.2329
#>  -0.3207  0.0970
#>   0.2285  0.2031
#> [ CPUFloatType{2,3,20} ]
#>