RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\text{ReLU}\) non-linearity to an input sequence.

nn_rnn(...)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional:
If TRUE, becomes a bidirectional RNN. Default: FALSE

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\text{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\text{input_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\text{hidden_size}\) Defaults to zero if not provided. where \(S=\text{num_layers} * \text{num_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\text{num_directions} * \text{hidden_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_lk: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_lk: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_lk: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_lk: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\text{hidden\_size}}\)

Examples

rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.4276  0.0772 -0.5042 -0.6518  0.1811 -0.8639  0.2111  0.4346 -0.8817
#>  -0.3943  0.1186  0.2027 -0.0821  0.2938  0.0389 -0.3633 -0.3283 -0.7592
#>  -0.1428 -0.5500  0.5352  0.4765  0.1696  0.7090  0.5303 -0.6220 -0.2569
#> 
#> Columns 10 to 18 -0.6619 -0.5207  0.8410 -0.3293  0.7054  0.0577 -0.8078 -0.5257  0.6800
#>  -0.3951 -0.6624  0.6272 -0.0957  0.3011 -0.6020  0.5136 -0.5943  0.4131
#>   0.6771  0.9127 -0.3022  0.0709 -0.4073 -0.3627 -0.7434  0.4078  0.6302
#> 
#> Columns 19 to 20 -0.4580 -0.0341
#>  -0.6723 -0.3000
#>   0.6360 -0.3711
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1384 -0.0091 -0.0056  0.3984  0.5643 -0.0460  0.8278  0.3454 -0.5000
#>   0.2556 -0.4035 -0.7108  0.3854 -0.3852  0.1034  0.5604  0.2520  0.1212
#>   0.3114 -0.4072  0.4441  0.6002 -0.2457 -0.1299  0.1112  0.3806 -0.3545
#> 
#> Columns 10 to 18 -0.2446  0.0692  0.4671 -0.5671  0.4301  0.1657  0.6823 -0.6196  0.3762
#>  -0.7497  0.0340 -0.3474 -0.6738 -0.0242  0.1581  0.5449 -0.4903 -0.3489
#>   0.1728  0.2955  0.3845 -0.4589  0.2551  0.4480 -0.4291 -0.4480  0.5182
#> 
#> Columns 19 to 20  0.0809 -0.4582
#>  -0.2338 -0.5071
#>  -0.1378 -0.0010
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3947 -0.4012 -0.3611  0.3505  0.0165  0.1398 -0.0617  0.0763  0.1940
#>   0.3219  0.4130  0.6590 -0.1642 -0.0500  0.5086 -0.3445 -0.0939 -0.2414
#>   0.3590 -0.1237  0.0934 -0.2235 -0.2663  0.2981  0.0762 -0.0298  0.4023
#> 
#> Columns 10 to 18  0.0585  0.5780 -0.4226 -0.5474 -0.1996 -0.0245  0.2509 -0.2825  0.0077
#>   0.5736  0.1683  0.3064 -0.3716  0.3403 -0.0382  0.5370 -0.4994  0.6753
#>  -0.3686  0.2479 -0.0329 -0.6391 -0.5372 -0.0355 -0.4228  0.1528  0.2182
#> 
#> Columns 19 to 20 -0.0757 -0.2216
#>  -0.5597  0.6125
#>  -0.1468 -0.1929
#> 
#> (4,.,.) = 
#>  Columns 1 to 9 -0.0795  0.1925  0.5405  0.0138 -0.4712 -0.0736 -0.4220  0.3339 -0.1666
#>  -0.0071 -0.2551 -0.1282 -0.0769  0.1891  0.4082  0.4852 -0.0548 -0.7814
#>   0.4382  0.2154  0.3009  0.3081 -0.6001  0.3156 -0.0491  0.0601  0.3104
#> 
#> Columns 10 to 18  0.8121  0.2617  0.6845 -0.3926  0.5111  0.3392  0.1697 -0.6972  0.5728
#>   0.1224 -0.1796  0.4606 -0.2305  0.0887  0.6502 -0.1052 -0.7079  0.3113
#>  -0.4494  0.2919  0.1243 -0.6592 -0.4256  0.2197  0.0171 -0.0662 -0.2708
#> 
#> Columns 19 to 20 -0.6355  0.4968
#>   0.2371 -0.2860
#>  -0.3225 -0.0697
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.0960 -0.3353 -0.4172 -0.5399 -0.2198  0.1494  0.0701 -0.1455  0.1152
#>   0.0498 -0.3871  0.1187  0.3674 -0.0330 -0.1328  0.3940 -0.0463 -0.2081
#>   0.3725  0.2881  0.6440  0.0512 -0.2188  0.4613  0.2503 -0.2195 -0.5595
#> 
#> Columns 10 to 18  0.0065 -0.1936  0.1489 -0.3465 -0.4079  0.2003 -0.4970 -0.1207 -0.1250
#>   0.2579  0.1203  0.1182 -0.4260  0.4201 -0.3926  0.0258 -0.2187 -0.0547
#>   0.1541  0.3313  0.5169 -0.1832 -0.0369  0.5818  0.1595 -0.6282  0.5641
#> 
#> Columns 19 to 20 -0.3074  0.1372
#>  -0.1573 -0.3119
#>  -0.3600  0.3056
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0959  0.0263 -0.4122 -0.0120 -0.2258 -0.5127  0.5458  0.1145 -0.0379
#>  -0.0000  0.1686 -0.2831  0.0969 -0.7582 -0.0024  0.4333 -0.6053  0.0907
#>   0.6877  0.3841  0.1194 -0.3983  0.5711  0.4428 -0.4015  0.2521  0.4396
#> 
#> Columns 10 to 18 -0.2907 -0.4035 -0.6876 -0.2241  0.4380  0.1664 -0.0698 -0.3933 -0.4121
#>  -0.3388 -0.2028 -0.5774 -0.1317 -0.3398 -0.3893  0.5495 -0.4529 -0.0072
#>   0.1676  0.2725  0.8899  0.6866  0.7163 -0.4537 -0.2172  0.1430 -0.0643
#> 
#> Columns 19 to 20 -0.6915 -0.2926
#>  -0.1140 -0.0890
#>   0.2944  0.4723
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0960 -0.3353 -0.4172 -0.5399 -0.2198  0.1494  0.0701 -0.1455  0.1152
#>   0.0498 -0.3871  0.1187  0.3674 -0.0330 -0.1328  0.3940 -0.0463 -0.2081
#>   0.3725  0.2881  0.6440  0.0512 -0.2188  0.4613  0.2503 -0.2195 -0.5595
#> 
#> Columns 10 to 18  0.0065 -0.1936  0.1489 -0.3465 -0.4079  0.2003 -0.4970 -0.1207 -0.1250
#>   0.2579  0.1203  0.1182 -0.4260  0.4201 -0.3926  0.0258 -0.2187 -0.0547
#>   0.1541  0.3313  0.5169 -0.1832 -0.0369  0.5818  0.1595 -0.6282  0.5641
#> 
#> Columns 19 to 20 -0.3074  0.1372
#>  -0.1573 -0.3119
#>  -0.3600  0.3056
#> [ CPUFloatType{2,3,20} ]
#>