RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7847  0.9186  0.2746  0.2146  0.4001 -0.4031  0.3518 -0.6876  0.6414
#>   0.4423  0.7451 -0.4374  0.2732  0.3404  0.1314  0.1832  0.3256  0.8725
#>  -0.9114  0.7255  0.5415 -0.2668  0.1125 -0.1175 -0.1738  0.4853 -0.1317
#> 
#> Columns 10 to 18  0.1028  0.3355 -0.3975  0.3840 -0.0445 -0.8893 -0.9370 -0.5425 -0.7912
#>   0.0478  0.7512 -0.7788 -0.7081 -0.2186 -0.5992 -0.8818  0.3562 -0.0703
#>  -0.7359  0.6260 -0.5624  0.4162  0.4805 -0.0774 -0.5281  0.8349  0.4023
#> 
#> Columns 19 to 20  0.6127 -0.7741
#>   0.7746 -0.0218
#>   0.1412 -0.9623
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5502  0.5415  0.1181  0.3207  0.1691 -0.0978  0.7220  0.4593 -0.2693
#>  -0.6192  0.4386  0.2981  0.4616  0.2439 -0.1955  0.7705 -0.0102 -0.1194
#>   0.0768  0.3511  0.0820 -0.1210  0.1155 -0.4469  0.5068 -0.0015  0.3423
#> 
#> Columns 10 to 18  0.2890 -0.2762 -0.1783 -0.5391 -0.1537 -0.6339 -0.3893 -0.5607 -0.3863
#>   0.4611  0.3649 -0.5377 -0.2108  0.1130 -0.7148 -0.3330  0.1479 -0.7573
#>   0.3023  0.2607 -0.3765  0.0356 -0.7024  0.1357 -0.3234  0.0120  0.7726
#> 
#> Columns 19 to 20 -0.0780 -0.0978
#>  -0.3810 -0.0329
#>  -0.5049  0.1115
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2294 -0.1180  0.6002 -0.0587 -0.0829 -0.4745  0.2941  0.1124  0.2578
#>  -0.1965  0.2757  0.3216 -0.3644 -0.1701  0.1780  0.3149 -0.1407  0.1991
#>  -0.4607  0.1050  0.4604 -0.4772  0.1782  0.0867  0.2383 -0.2436  0.0989
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7224  0.5578 -0.1584  0.8061  0.5589  0.7441  0.6438 -0.7797  0.1906
#>  -0.2246 -0.0483  0.3602  0.0276  0.7001  0.7932  0.2435 -0.1061 -0.2843
#>   0.0943  0.1658  0.5184 -0.1662  0.4578  0.3031  0.4837  0.5420  0.8294
#> 
#> Columns 10 to 18 -0.1042 -0.5802 -0.5159  0.5630  0.4544 -0.4228 -0.8154  0.0645 -0.6682
#>   0.1260 -0.5953  0.0879  0.6505  0.1827  0.0325 -0.5092 -0.1255 -0.5005
#>   0.6180 -0.0866  0.1659 -0.2526 -0.2885 -0.2210 -0.3336  0.6744 -0.4275
#> 
#> Columns 19 to 20  0.0141  0.2933
#>   0.1116 -0.5053
#>  -0.0049  0.0725
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4204 -0.2382 -0.1680 -0.3013  0.2785  0.3223  0.0291  0.3430  0.2877
#>  -0.2758  0.0410 -0.0476 -0.2510 -0.0876 -0.1293  0.3132  0.1082  0.3376
#>  -0.3898 -0.0496  0.5339 -0.1973  0.4080 -0.2912  0.2859  0.3656 -0.3110
#> 
#> Columns 10 to 18 -0.5595  0.2146 -0.4549 -0.1014  0.4855  0.0071  0.1372  0.2667  0.1005
#>  -0.3453  0.3452 -0.2122 -0.2586  0.1409 -0.1420  0.0416  0.0880  0.0484
#>  -0.3970 -0.2056 -0.1465 -0.3860 -0.1983 -0.1673  0.0050 -0.1231 -0.1219
#> 
#> Columns 19 to 20  0.2945  0.2170
#>   0.3455 -0.0559
#>   0.1705 -0.3733
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>