RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8752  0.3649 -0.4802 -0.5042 -0.4120 -0.0699  0.3109 -0.8710 -0.1498
#>  -0.3455  0.4604 -0.0027 -0.6833  0.4235 -0.1267  0.5774 -0.4593 -0.5573
#>   0.5108  0.3109 -0.0721  0.4710 -0.3885 -0.2080 -0.0101 -0.4080  0.4942
#> 
#> Columns 10 to 18  0.5406  0.2444 -0.3208 -0.4569 -0.5747  0.6216  0.1119 -0.4619  0.2788
#>   0.1809  0.7756 -0.8180  0.4955  0.1428  0.4588  0.8033  0.7239 -0.5687
#>   0.8659  0.5625 -0.2521  0.6767  0.0587 -0.5796 -0.4738  0.0197 -0.0726
#> 
#> Columns 19 to 20  0.7634 -0.3578
#>   0.5932 -0.2177
#>  -0.4568  0.8086
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1352 -0.8735 -0.4096 -0.1874  0.1024  0.3761 -0.3022 -0.1492  0.4086
#>  -0.0434 -0.4862 -0.4686 -0.6752 -0.1225  0.0406  0.0103  0.0964 -0.0799
#>  -0.5015 -0.3762 -0.1206  0.0632  0.0661  0.6411  0.0671  0.6508 -0.4023
#> 
#> Columns 10 to 18  0.2807 -0.0419  0.0684  0.0460 -0.0094  0.2243 -0.2616 -0.5091  0.4600
#>   0.2270 -0.4878 -0.0032  0.2887  0.3299  0.1648  0.2672  0.1405  0.3632
#>   0.3293 -0.2897 -0.3647  0.1226 -0.5316  0.1045  0.2427  0.5972 -0.1997
#> 
#> Columns 19 to 20  0.4631  0.5856
#>   0.5335  0.7384
#>   0.2689  0.7917
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2607 -0.4368 -0.1555 -0.2233 -0.3751  0.0421 -0.4570  0.0051  0.4373
#>  -0.2021 -0.4024  0.1945 -0.3751 -0.2542 -0.1189 -0.2959  0.0864  0.2498
#>   0.1363 -0.3983 -0.2644 -0.0910  0.1752  0.1150 -0.6801  0.0700 -0.3682
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0040 -0.0242  0.0965 -0.4691  0.0987 -0.5186  0.2165 -0.0967 -0.3461
#>   0.7915  0.3989  0.1860 -0.2071  0.3040 -0.5263 -0.2556  0.4140 -0.6934
#>   0.0480  0.0997  0.5306 -0.0117 -0.8375 -0.7597  0.4401  0.6836  0.4976
#> 
#> Columns 10 to 18  0.5711  0.6262  0.3444 -0.1831  0.5318  0.6369 -0.0566 -0.2895  0.5062
#>   0.8526  0.7367 -0.1123  0.2737  0.3890 -0.2682  0.6238  0.1635  0.6859
#>  -0.4734  0.1357 -0.1212  0.1345  0.7609  0.0479 -0.1564  0.7124 -0.6227
#> 
#> Columns 19 to 20 -0.1033 -0.1024
#>  -0.0110 -0.5992
#>  -0.0739  0.0532
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1685 -0.3496 -0.1529 -0.4178 -0.3470  0.0533 -0.3084 -0.1386 -0.0018
#>   0.2658 -0.1400 -0.1866 -0.5994 -0.1384 -0.1422 -0.5716  0.0345 -0.2763
#>  -0.0630 -0.1792 -0.3406 -0.7145 -0.1127  0.0751  0.0010 -0.2100  0.6095
#> 
#> Columns 10 to 18  0.0286 -0.1742 -0.1719  0.3886 -0.4741  0.4250  0.4416 -0.1660 -0.1619
#>  -0.3017 -0.0118 -0.3989  0.7429 -0.1968  0.1308  0.3880  0.3407 -0.1040
#>  -0.0878 -0.0111  0.0619  0.4489 -0.1777 -0.3036  0.2135 -0.5143  0.4088
#> 
#> Columns 19 to 20  0.0327  0.5722
#>   0.4525  0.3924
#>  -0.3012  0.4614
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>