RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8632  0.2606 -0.7561  0.3238  0.5758 -0.9295  0.6206  0.7773  0.0156
#>  -0.4929  0.2741 -0.9271  0.6827  0.4148 -0.9234 -0.3215  0.8569 -0.3811
#>  -0.1974  0.1782 -0.4820  0.5614 -0.1610  0.0945  0.0546  0.7993 -0.6836
#> 
#> Columns 10 to 18  0.0800  0.8516 -0.8327 -0.7442  0.8087 -0.4025  0.7856 -0.5152  0.1134
#>  -0.0770 -0.5458 -0.7040 -0.4072  0.4399  0.7389  0.7642  0.6092  0.6220
#>  -0.6017 -0.1184 -0.8335  0.3620 -0.4830 -0.6192  0.0020  0.3373  0.5506
#> 
#> Columns 19 to 20 -0.1426 -0.6890
#>  -0.0678 -0.9730
#>  -0.3883  0.7394
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2283  0.3322 -0.2536 -0.0099  0.3812  0.4139  0.2319  0.5683  0.0016
#>   0.4454  0.4583 -0.6253 -0.3262  0.3128  0.5423  0.2797  0.0986  0.1413
#>  -0.4787  0.2650 -0.2538 -0.4812 -0.0043  0.1951 -0.1059  0.2570 -0.3133
#> 
#> Columns 10 to 18  0.5097 -0.3602 -0.0984 -0.0912 -0.3948  0.4190  0.1389 -0.3516  0.0742
#>   0.6472 -0.5922 -0.1273 -0.0898 -0.2280  0.4313 -0.0195 -0.3458 -0.3630
#>  -0.2426  0.1201 -0.1422  0.1834 -0.0601  0.1023 -0.2650 -0.3652 -0.0128
#> 
#> Columns 19 to 20 -0.0921 -0.7159
#>   0.1415 -0.3896
#>   0.0574  0.2064
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2523  0.0743 -0.2917  0.0008  0.0287  0.3772 -0.1288 -0.0446  0.1969
#>  -0.1850 -0.2193  0.3849 -0.3458 -0.6648  0.0691 -0.0999 -0.1181 -0.3041
#>  -0.0350  0.2601  0.0828 -0.3278  0.0189 -0.1383 -0.2389 -0.1243 -0.0835
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5805 -0.0152  0.3949 -0.2913 -0.6923  0.4125  0.3901  0.6397  0.2422
#>  -0.2745 -0.0733  0.3335 -0.8813  0.1085 -0.2381 -0.3142 -0.2977  0.6824
#>  -0.5595  0.5570  0.2514  0.4974  0.2614  0.0057  0.5995  0.3866  0.6004
#> 
#> Columns 10 to 18  0.7074  0.4904  0.5232 -0.2944  0.2058 -0.8497 -0.2329 -0.8124 -0.2780
#>  -0.5115  0.4549  0.2427  0.4174  0.0823  0.0596 -0.0900  0.5328  0.1347
#>  -0.0971 -0.1035  0.8416 -0.2325 -0.3694 -0.2634  0.8250  0.2056  0.8689
#> 
#> Columns 19 to 20  0.3710 -0.7871
#>   0.2298  0.1624
#>  -0.1423 -0.5529
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2843  0.2757  0.3211  0.2911 -0.5787 -0.1491  0.2332 -0.0854 -0.4842
#>   0.3274  0.0051 -0.0463 -0.3501 -0.3122 -0.0225  0.0954 -0.0887 -0.2178
#>   0.3806  0.5730 -0.2553  0.2304  0.0344 -0.3198  0.5400  0.1179 -0.1098
#> 
#> Columns 10 to 18  0.0873  0.0313  0.2982 -0.1643  0.0681 -0.5363  0.2195  0.0500  0.4808
#>  -0.0363 -0.6194 -0.1544 -0.1873  0.2113  0.0305 -0.1097  0.0286  0.0810
#>   0.2071 -0.5172 -0.0095 -0.4391  0.0711 -0.5877 -0.0729 -0.0804 -0.1229
#> 
#> Columns 19 to 20 -0.1740 -0.2687
#>  -0.4186 -0.1769
#>  -0.0232 -0.5438
#> [ CPUFloatType{2,3,20} ]
#>