RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7434  0.2844  0.0016  0.8185  0.6531  0.4408  0.2513  0.4043  0.1344
#>  -0.4084 -0.0432  0.0035 -0.0020  0.3743 -0.1715 -0.0055 -0.1983  0.3436
#>  -0.7834  0.3324  0.1143 -0.8157 -0.1834 -0.5303 -0.1466 -0.3705  0.5860
#> 
#> Columns 10 to 18 -0.2941 -0.0811 -0.0655 -0.0560 -0.0679 -0.5190  0.4709  0.8197 -0.3277
#>   0.1659 -0.2903 -0.3240  0.2294  0.8071  0.6459 -0.6420  0.1741 -0.1008
#>   0.7103  0.2403 -0.2498  0.5758  0.9233 -0.2942 -0.9029  0.7517  0.3229
#> 
#> Columns 19 to 20 -0.5622  0.1777
#>  -0.4280  0.4360
#>   0.3474  0.4417
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3215  0.5515  0.3459  0.2112  0.1356 -0.1766  0.1357  0.1387  0.4026
#>  -0.1416  0.3790  0.1404  0.2297 -0.6452  0.1865 -0.4193 -0.0407 -0.3230
#>   0.2298  0.0540 -0.0554  0.0693 -0.2164 -0.0604 -0.1563  0.3245 -0.3044
#> 
#> Columns 10 to 18  0.6690  0.2629  0.4034 -0.2537  0.4071 -0.0820 -0.0872 -0.0918 -0.0751
#>   0.4012  0.0271 -0.0440 -0.2764  0.4863 -0.0607  0.0546  0.1313  0.1613
#>   0.5691  0.5713 -0.4025 -0.3613  0.3051  0.2366  0.2158  0.4308 -0.3279
#> 
#> Columns 19 to 20 -0.0694  0.2273
#>  -0.3657 -0.0064
#>   0.4104  0.4091
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3717  0.1676  0.0138  0.5411 -0.4264  0.0728  0.2278  0.0721 -0.0395
#>  -0.2702  0.0194  0.3398  0.2389 -0.5255 -0.3706 -0.0438  0.2220 -0.1981
#>  -0.4718  0.1189 -0.1507  0.6715 -0.0228  0.3978  0.0105  0.2033 -0.5316
#> 
#> Columns 10 to 18  0.6726  0.0628 -0.3861 -0.5772  0.5141 -0.0998 -0.3381  0.3791  0.0948
#>   0.4351 -0.1256 -0.1891 -0.3737 -0.1287  0.2312  0.1125 -0.0475  0.6847
#>   0.1752 -0.0895 -0.2410 -0.6497  0.4282  0.1321  0.1799  0.3171  0.7603
#> 
#> Columns 19 to 20 -0.0338  0.0285
#>  -0.0107 -0.0276
#>  -0.0632  0.2465
#> 
#> (4,.,.) = 
#>  Columns 1 to 9 -0.4956  0.2495  0.1938  0.6142 -0.7948  0.2498 -0.2469 -0.3769  0.1721
#>  -0.0427 -0.0137  0.4365  0.4746 -0.4662 -0.2331 -0.4956  0.2313  0.0292
#>  -0.6142  0.3817  0.4657  0.3337 -0.7294  0.2089 -0.3426 -0.3479 -0.1028
#> 
#> Columns 10 to 18  0.2498 -0.2332 -0.2562 -0.4496  0.2118 -0.3066 -0.3265  0.5211  0.2567
#>   0.6424 -0.2713 -0.3379 -0.2258 -0.2887 -0.0166 -0.3823  0.0039  0.7031
#>   0.4339 -0.5604 -0.1774  0.2026  0.0930 -0.3116 -0.1249  0.4010  0.2518
#> 
#> Columns 19 to 20 -0.0292 -0.0534
#>  -0.1144 -0.1243
#>  -0.1189 -0.0810
#> 
#> (5,.,.) = 
#>  Columns 1 to 9 -0.0346  0.2016  0.3038 -0.2053 -0.6229 -0.1266 -0.5756 -0.0109  0.0019
#>   0.0295 -0.0835  0.3223  0.3028 -0.4772 -0.3311 -0.6521 -0.0252 -0.2980
#>  -0.0244  0.0057  0.2770 -0.0663 -0.2242 -0.3654 -0.6100  0.1517 -0.1629
#> 
#> Columns 10 to 18  0.6330  0.0211 -0.0987 -0.1026  0.1900 -0.0459 -0.1214  0.3469  0.4157
#>   0.6977 -0.1733 -0.2939 -0.4534  0.1784  0.1775 -0.5701 -0.0754  0.6362
#>   0.6684  0.0042 -0.0156 -0.1654  0.1346  0.4302  0.0809 -0.2777  0.6461
#> 
#> Columns 19 to 20  0.1809 -0.1792
#>  -0.2223  0.1212
#>   0.1024  0.1324
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.2880  0.2445  0.1000 -0.0137 -0.2287  0.2212 -0.4598 -0.1706 -0.4999
#>   0.2428 -0.0218 -0.7798  0.3126 -0.5297 -0.1007 -0.6938 -0.3113 -0.7709
#>   0.4633  0.0818 -0.6957 -0.5623 -0.5963  0.3772 -0.9365 -0.0530 -0.4626
#> 
#> Columns 10 to 18 -0.5619 -0.1573  0.0434  0.5007  0.2885 -0.3552  0.7300 -0.3641  0.0935
#>  -0.1233  0.0221  0.6379 -0.4574  0.4790  0.3875 -0.3689 -0.6219  0.0886
#>  -0.6095 -0.6709  0.0117 -0.1393  0.7393  0.6836  0.4758 -0.2844 -0.0577
#> 
#> Columns 19 to 20  0.1469  0.5601
#>   0.0360 -0.6894
#>  -0.1248  0.4876
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0346  0.2016  0.3038 -0.2053 -0.6229 -0.1266 -0.5756 -0.0109  0.0019
#>   0.0295 -0.0835  0.3223  0.3028 -0.4772 -0.3311 -0.6521 -0.0252 -0.2980
#>  -0.0244  0.0057  0.2770 -0.0663 -0.2242 -0.3654 -0.6100  0.1517 -0.1629
#> 
#> Columns 10 to 18  0.6330  0.0211 -0.0987 -0.1026  0.1900 -0.0459 -0.1214  0.3469  0.4157
#>   0.6977 -0.1733 -0.2939 -0.4534  0.1784  0.1775 -0.5701 -0.0754  0.6362
#>   0.6684  0.0042 -0.0156 -0.1654  0.1346  0.4302  0.0809 -0.2777  0.6461
#> 
#> Columns 19 to 20  0.1809 -0.1792
#>  -0.2223  0.1212
#>   0.1024  0.1324
#> [ CPUFloatType{2,3,20} ]
#>