RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2790  0.4557  0.0415  0.8348 -0.7888 -0.6704  0.5190  0.8056 -0.0735
#>  -0.3734  0.7112 -0.4892  0.3629 -0.8767 -0.2695 -0.6813 -0.1264  0.4994
#>  -0.1336  0.3137 -0.2728  0.3881 -0.9106  0.2381  0.6806  0.5845  0.0070
#> 
#> Columns 10 to 18  0.6130 -0.5999  0.7299 -0.0331 -0.0195 -0.2805 -0.0049 -0.1615  0.5580
#>   0.4963 -0.8276  0.3172 -0.2244  0.2435 -0.6658  0.2840 -0.7537  0.6139
#>  -0.6340 -0.8699  0.4936  0.2009  0.5499 -0.8105 -0.5179 -0.2929  0.3923
#> 
#> Columns 19 to 20  0.2897  0.6919
#>  -0.4266 -0.3322
#>  -0.6059  0.7904
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1542  0.4577  0.2647 -0.4398 -0.6971 -0.1115 -0.5807  0.2753 -0.6948
#>  -0.2141  0.1742 -0.4470 -0.0062 -0.5819 -0.2804  0.0925  0.2187 -0.2401
#>  -0.1620  0.2789  0.3353 -0.2993 -0.6950 -0.4902 -0.0942 -0.1694 -0.2142
#> 
#> Columns 10 to 18  0.4217 -0.6584 -0.4289  0.5238 -0.7257 -0.0498 -0.3210  0.2025  0.2758
#>  -0.4385 -0.8835  0.1600  0.4939 -0.5755 -0.4960 -0.2327 -0.8242  0.2941
#>  -0.1414 -0.7278 -0.6218  0.2732 -0.4391 -0.3292 -0.2274 -0.1094  0.5159
#> 
#> Columns 19 to 20  0.5948 -0.0022
#>   0.2126  0.0036
#>   0.4086 -0.4942
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5362  0.1198 -0.3093 -0.5701 -0.2285 -0.2572  0.3774  0.0206  0.0784
#>  -0.1415 -0.0714  0.1182 -0.6636 -0.7014  0.6002 -0.1912  0.4512 -0.5636
#>   0.4931  0.4051 -0.4790 -0.5175 -0.0738  0.5286 -0.5998 -0.0225 -0.6865
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2258  0.4351 -0.6280 -0.5593 -0.6730  0.0394 -0.2843  0.1644  0.5445
#>  -0.0109 -0.0469 -0.3889  0.3547 -0.0988 -0.4007 -0.0149 -0.0693 -0.6475
#>  -0.1863 -0.5970 -0.1650 -0.1984  0.1591  0.4999 -0.4886  0.7543 -0.1315
#> 
#> Columns 10 to 18 -0.5367  0.5344 -0.0116  0.2107  0.2568 -0.7885 -0.3532  0.6266  0.0290
#>  -0.3111 -0.0725 -0.0972  0.3587 -0.5783 -0.0501  0.0016 -0.2265  0.2206
#>  -0.5436  0.7369 -0.5528  0.6331 -0.7509 -0.1326  0.0966  0.6476 -0.7396
#> 
#> Columns 19 to 20 -0.6417  0.2754
#>  -0.0371 -0.4236
#>  -0.4774 -0.5815
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2077 -0.1160  0.1590  0.4533 -0.5216 -0.1721  0.2468  0.6150  0.0826
#>   0.5147  0.5231 -0.0738 -0.4369 -0.3332 -0.1699  0.0088 -0.2080 -0.1336
#>   0.3555  0.2137 -0.1071  0.0185 -0.4082  0.3647  0.0914  0.5954  0.0644
#> 
#> Columns 10 to 18  0.1633 -0.1852  0.1338  0.2282  0.2983 -0.7948 -0.0636 -0.4080  0.5485
#>   0.2542 -0.1958 -0.6295  0.0467 -0.5483 -0.4252 -0.2614  0.1678 -0.2035
#>  -0.1077 -0.2211  0.0697  0.1194  0.2787 -0.5930 -0.8432  0.5588 -0.4302
#> 
#> Columns 19 to 20 -0.2818 -0.1806
#>   0.1082 -0.1443
#>  -0.5446  0.5639
#> [ CPUFloatType{2,3,20} ]
#>