RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.0589  0.6076  0.5733  0.4325 -0.7804 -0.5884 -0.7990 -0.5201 -0.5453
#>  -0.3423 -0.1028  0.8203 -0.4395 -0.8638 -0.2453 -0.7419 -0.5734 -0.9229
#>  -0.0001  0.2485 -0.1473  0.5797 -0.1626 -0.1213 -0.3884 -0.5131  0.2432
#> 
#> Columns 10 to 18  0.2049  0.6141  0.0159 -0.5202 -0.0390  0.8628  0.0559 -0.6653  0.5929
#>   0.1484 -0.3955  0.4318 -0.6923  0.6768  0.9427 -0.5935 -0.1124  0.8008
#>  -0.1115  0.3922 -0.0407  0.7891 -0.1390  0.3705 -0.2486  0.8745 -0.3595
#> 
#> Columns 19 to 20  0.0777 -0.9009
#>  -0.0200 -0.2583
#>  -0.1381 -0.4595
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5747 -0.2374  0.3985  0.6116 -0.7816  0.0973 -0.2544 -0.2252 -0.2584
#>  -0.2011 -0.3030  0.4559  0.2132 -0.6755  0.1609 -0.8239 -0.2612 -0.6095
#>  -0.5134  0.3075 -0.2427  0.1875 -0.3368  0.1697 -0.3024 -0.5115  0.0946
#> 
#> Columns 10 to 18  0.5029 -0.4663 -0.6517 -0.3916 -0.3694 -0.3079  0.0697  0.4641  0.4321
#>   0.8177  0.0184 -0.5879 -0.3717 -0.0954  0.1382 -0.1993  0.5621 -0.1939
#>   0.3419 -0.0541  0.0665  0.0764  0.6821  0.0040  0.0504  0.6075  0.0703
#> 
#> Columns 19 to 20 -0.1393 -0.0169
#>  -0.2283  0.0531
#>  -0.0904  0.2892
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4901  0.5564  0.4157 -0.1239 -0.3616 -0.3069 -0.7718  0.2739  0.1213
#>   0.4148  0.1776  0.2237  0.0323 -0.5166 -0.3952 -0.6464  0.1560  0.1433
#>   0.3964  0.0038  0.3253  0.0715  0.1306 -0.0327 -0.2969  0.1245 -0.0192
#> 
#> Columns 10 to 18  0.7138 -0.0042  0.1386 -0.0459 -0.4418  0.2562 -0.1983  0.0256  0.4770
#>   0.7310  0.4867 -0.2056 -0.4410 -0.1978  0.4279 -0.1295  0.2951  0.4745
#>   0.6958  0.5029  0.0787 -0.2960  0.3043  0.3281 -0.3449  0.3081  0.4197
#> 
#> Columns 19 to 20  0.1745 -0.1350
#>   0.0668 -0.4015
#>   0.5455  0.3373
#> 
#> (4,.,.) = 
#>  Columns 1 to 9 -0.1242 -0.1825  0.0703  0.3368 -0.1879 -0.6188 -0.3606  0.2400 -0.4979
#>  -0.2285 -0.1238  0.2816  0.7158 -0.3795 -0.1800 -0.0903  0.0951 -0.4055
#>  -0.4161  0.1562  0.1320  0.5917 -0.2003  0.5784 -0.1302 -0.1113 -0.7354
#> 
#> Columns 10 to 18  0.6470  0.0346 -0.6063 -0.0664  0.3145  0.3909 -0.3163  0.0040  0.6242
#>   0.3401 -0.2107 -0.5455 -0.4844 -0.0769  0.0620 -0.1112  0.5133  0.6606
#>   0.4108 -0.3172 -0.3713 -0.5535  0.3277  0.0108 -0.0800  0.5410  0.1190
#> 
#> Columns 19 to 20 -0.1511 -0.0390
#>   0.2056  0.1060
#>   0.1649  0.4299
#> 
#> (5,.,.) = 
#>  Columns 1 to 9 -0.0957  0.3769  0.1599  0.2209 -0.5452 -0.2272 -0.5200 -0.0675 -0.3636
#>   0.3314  0.3243  0.5617  0.3386 -0.1621  0.4163 -0.6332 -0.0253 -0.0027
#>   0.1690  0.3272  0.5195 -0.1727 -0.1203 -0.0701 -0.5596 -0.2452  0.1866
#> 
#> Columns 10 to 18  0.7646  0.0595 -0.1395 -0.1890  0.1511  0.3273 -0.2851  0.1429  0.4489
#>   0.7122  0.0548 -0.1359 -0.2292 -0.1363 -0.1930 -0.3282  0.5265  0.3602
#>   0.7579  0.1487  0.1720 -0.0712 -0.0081  0.3472 -0.1236  0.3269  0.3627
#> 
#> Columns 19 to 20 -0.3324  0.0199
#>   0.4319  0.0719
#>   0.0541 -0.1996
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0859  0.2441 -0.3785  0.4773 -0.6194  0.7348 -0.2389  0.3932 -0.5751
#>  -0.3461 -0.4299 -0.5382 -0.0556  0.2770 -0.4997  0.3906 -0.4267 -0.2962
#>   0.0901 -0.0487 -0.2221  0.7402 -0.3091  0.1068  0.2309  0.0844 -0.8371
#> 
#> Columns 10 to 18 -0.7465 -0.0237 -0.5216  0.2091  0.7810  0.0142 -0.0300  0.0860 -0.4565
#>   0.4484 -0.2549  0.6987 -0.4158 -0.6085 -0.0974  0.0892 -0.3520  0.0667
#>  -0.1097 -0.0399  0.2584 -0.1982  0.0520 -0.0103 -0.0936  0.3060  0.0546
#> 
#> Columns 19 to 20  0.1902  0.0056
#>   0.3507  0.1104
#>  -0.0194  0.3957
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0957  0.3769  0.1599  0.2209 -0.5452 -0.2272 -0.5200 -0.0675 -0.3636
#>   0.3314  0.3243  0.5617  0.3386 -0.1621  0.4163 -0.6332 -0.0253 -0.0027
#>   0.1690  0.3272  0.5195 -0.1727 -0.1203 -0.0701 -0.5596 -0.2452  0.1866
#> 
#> Columns 10 to 18  0.7646  0.0595 -0.1395 -0.1890  0.1511  0.3273 -0.2851  0.1429  0.4489
#>   0.7122  0.0548 -0.1359 -0.2292 -0.1363 -0.1930 -0.3282  0.5265  0.3602
#>   0.7579  0.1487  0.1720 -0.0712 -0.0081  0.3472 -0.1236  0.3269  0.3627
#> 
#> Columns 19 to 20 -0.3324  0.0199
#>   0.4319  0.0719
#>   0.0541 -0.1996
#> [ CPUFloatType{2,3,20} ]
#>