RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\text{ReLU}\) non-linearity to an input sequence.

nn_rnn(...)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional:
If TRUE, becomes a bidirectional RNN. Default: FALSE

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\text{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\text{input_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\text{hidden_size}\) Defaults to zero if not provided. where \(S=\text{num_layers} * \text{num_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\text{num_directions} * \text{hidden_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_lk: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_lk: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_lk: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_lk: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\text{hidden\_size}}\)

Examples

rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.7935  0.2718  0.1289  0.1707  0.1594  0.0460 -0.1383  0.1297  0.5121
#>  -0.4831 -0.3777 -0.4205 -0.2305 -0.5965  0.3670 -0.1254 -0.3236 -0.4388
#>   0.6608  0.2711 -0.4220  0.4028  0.0329  0.4820 -0.5917 -0.6846  0.5982
#> 
#> Columns 10 to 18  0.3861  0.5258  0.9756  0.2645 -0.8347 -0.3985  0.6260 -0.1853  0.0500
#>   0.5284  0.3175  0.4477  0.7648 -0.8956  0.4296 -0.4579 -0.5839 -0.7793
#>   0.3974 -0.0316 -0.1947  0.3754 -0.4560 -0.3069 -0.7622 -0.1921  0.3572
#> 
#> Columns 19 to 20 -0.7018 -0.5608
#>  -0.8330  0.1409
#>  -0.5656 -0.2236
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4525 -0.6323 -0.3412  0.3695 -0.1300 -0.1674  0.2074 -0.3007 -0.7431
#>   0.3558  0.0356 -0.1873 -0.2204  0.4117  0.0736 -0.3915  0.1479  0.1560
#>   0.2560 -0.6116 -0.5343  0.6806  0.5847 -0.4672  0.0175 -0.4217 -0.5562
#> 
#> Columns 10 to 18  0.2527  0.3553  0.1955  0.2906 -0.2088  0.1461  0.2169 -0.4240 -0.5750
#>  -0.0075  0.3391  0.6301 -0.0882  0.0180 -0.4382 -0.4742  0.2710 -0.3647
#>  -0.1360  0.7039  0.6356 -0.0126 -0.0170  0.2918 -0.2864  0.0849 -0.1393
#> 
#> Columns 19 to 20  0.2195 -0.6103
#>  -0.1894 -0.3396
#>   0.1971 -0.3188
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5165 -0.1023 -0.3586  0.5728  0.4130 -0.0611 -0.0012  0.1252 -0.1688
#>   0.3854  0.4082 -0.2519  0.6531  0.2384 -0.4087  0.2169  0.0199  0.1643
#>   0.2344  0.3692 -0.3089  0.1477 -0.0045 -0.3425  0.0415  0.1594  0.1027
#> 
#> Columns 10 to 18 -0.0677 -0.0156  0.2411  0.1077 -0.0493 -0.4256 -0.5617 -0.0536 -0.4252
#>  -0.1188  0.2287  0.1779  0.1863  0.0187 -0.4115 -0.5192 -0.5680  0.1603
#>  -0.0829  0.0477 -0.1067  0.2451  0.0462 -0.0636 -0.7689  0.2251  0.1473
#> 
#> Columns 19 to 20  0.1399  0.0634
#>   0.1258  0.0538
#>   0.2013 -0.1329
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.4108 -0.0364 -0.5829  0.5528  0.6286 -0.1860  0.0602  0.1608  0.0767
#>   0.4323 -0.4556 -0.4257  0.7669  0.4272 -0.3919  0.3878 -0.0905 -0.4962
#>   0.6181  0.2465 -0.0297  0.6352  0.4836 -0.6454  0.3771  0.2815  0.1804
#> 
#> Columns 10 to 18 -0.2420  0.0618 -0.0203 -0.2064 -0.3554 -0.3258 -0.6814 -0.3385 -0.1409
#>  -0.2005  0.4191  0.2829 -0.1281 -0.0563 -0.1578 -0.4407 -0.2603 -0.1883
#>   0.1731  0.3292 -0.0186  0.0225 -0.1823 -0.4862 -0.6698 -0.1850  0.2304
#> 
#> Columns 19 to 20  0.1453 -0.0964
#>   0.1059 -0.2024
#>   0.0721  0.3804
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.2013 -0.0109 -0.3759  0.6971  0.5140 -0.3366  0.0379  0.1815 -0.4606
#>   0.3705  0.2135 -0.4904  0.6381  0.2702 -0.0942  0.2682  0.1775 -0.0991
#>   0.5582 -0.4385 -0.4098  0.8392  0.4007  0.1442  0.4666 -0.0339 -0.2426
#> 
#> Columns 10 to 18  0.0246  0.3091  0.0974 -0.2843 -0.5269 -0.5638 -0.4669 -0.4264 -0.3810
#>  -0.0817 -0.0342  0.4482  0.1686 -0.1223 -0.4237 -0.4241 -0.2176 -0.0196
#>   0.0178  0.1346  0.6702  0.2642  0.2793  0.0240 -0.3416 -0.3356 -0.1683
#> 
#> Columns 19 to 20  0.1584  0.0582
#>   0.1965 -0.1163
#>   0.2543 -0.3745
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.0163 -0.3121 -0.7693 -0.5169 -0.0080 -0.6754 -0.3799  0.3591  0.3520
#>  -0.4029 -0.1816 -0.3482  0.1835 -0.1503 -0.1366 -0.5040  0.3133 -0.1304
#>  -0.6669  0.6148 -0.1893  0.3126  0.5343  0.2303 -0.2049 -0.2356 -0.3536
#> 
#> Columns 10 to 18 -0.0814  0.0721  0.7363  0.6030 -0.5891  0.5513  0.0066  0.4777 -0.2228
#>  -0.0926  0.5105  0.3221  0.4246  0.5064  0.0451  0.2407 -0.2884 -0.1524
#>  -0.7204  0.6738  0.3283 -0.4297  0.8138 -0.7036  0.1409 -0.5870 -0.6643
#> 
#> Columns 19 to 20 -0.3675  0.7511
#>  -0.4025  0.4895
#>   0.0948  0.6117
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2013 -0.0109 -0.3759  0.6971  0.5140 -0.3366  0.0379  0.1815 -0.4606
#>   0.3705  0.2135 -0.4904  0.6381  0.2702 -0.0942  0.2682  0.1775 -0.0991
#>   0.5582 -0.4385 -0.4098  0.8392  0.4007  0.1442  0.4666 -0.0339 -0.2426
#> 
#> Columns 10 to 18  0.0246  0.3091  0.0974 -0.2843 -0.5269 -0.5638 -0.4669 -0.4264 -0.3810
#>  -0.0817 -0.0342  0.4482  0.1686 -0.1223 -0.4237 -0.4241 -0.2176 -0.0196
#>   0.0178  0.1346  0.6702  0.2642  0.2793  0.0240 -0.3416 -0.3356 -0.1683
#> 
#> Columns 19 to 20  0.1584  0.0582
#>   0.1965 -0.1163
#>   0.2543 -0.3745
#> [ CPUFloatType{2,3,20} ]
#>