RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0735 -0.0964 -0.5002 -0.6189  0.1166 -0.6215 -0.8296  0.5669 -0.5907
#>   0.7463  0.2492 -0.7027  0.6216 -0.5792 -0.8808 -0.4556 -0.5488  0.5506
#>   0.5688 -0.1357 -0.3157 -0.4171  0.6697  0.1366  0.4252  0.4776  0.4132
#> 
#> Columns 10 to 18  0.3799 -0.3553  0.3740 -0.7497 -0.2430  0.1731 -0.0252 -0.1681  0.3415
#>  -0.1455 -0.6224 -0.3171  0.3219  0.6010  0.3198  0.4112 -0.2396  0.4932
#>  -0.0957 -0.0726 -0.1531  0.6399  0.5241 -0.3840 -0.6945 -0.4487 -0.4888
#> 
#> Columns 19 to 20  0.1132  0.1269
#>   0.6614 -0.6572
#>   0.5792 -0.5615
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3076  0.5820 -0.6326 -0.0410  0.6036  0.2979 -0.4566 -0.0942 -0.2985
#>   0.3530  0.6217  0.4498  0.1639 -0.2356 -0.0310 -0.7001  0.6124 -0.4254
#>   0.6490  0.1042 -0.5253 -0.1980 -0.5328  0.2430 -0.0389  0.3054 -0.2851
#> 
#> Columns 10 to 18 -0.3028 -0.1716  0.1475 -0.3526 -0.3018  0.1193 -0.2508 -0.2569  0.1384
#>  -0.1928  0.1336  0.3672  0.0985 -0.6056 -0.7685 -0.7737 -0.2484 -0.5200
#>   0.1782 -0.3215  0.5119  0.2708  0.3380 -0.1169  0.3840  0.2634  0.0033
#> 
#> Columns 19 to 20  0.2030  0.0536
#>  -0.0769 -0.0059
#>   0.6759 -0.5978
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4914  0.3892 -0.4414 -0.0458  0.2824  0.0200 -0.5675  0.1372 -0.2072
#>   0.4976  0.3099 -0.7304  0.0143  0.4308  0.1640 -0.4122 -0.3054  0.4663
#>   0.2970  0.3431  0.0070  0.2858 -0.3449 -0.5582 -0.5764  0.0440 -0.3896
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1894 -0.0244  0.4809  0.2909 -0.1930  0.0707 -0.5971 -0.4189  0.2074
#>   0.4794  0.4023  0.8919  0.0482 -0.2296  0.6777 -0.2501 -0.4585  0.1033
#>   0.6803 -0.3585  0.0285 -0.5924  0.2983 -0.0663 -0.2627 -0.3797 -0.2031
#> 
#> Columns 10 to 18  0.2639  0.4342 -0.6026 -0.4516  0.1497 -0.0998  0.2703  0.6502 -0.4032
#>   0.4010  0.5177 -0.5838 -0.7930  0.5604 -0.5662 -0.3298  0.2858 -0.3453
#>   0.4179 -0.1651 -0.3406 -0.7556 -0.3173 -0.6492  0.3546  0.3490 -0.2095
#> 
#> Columns 19 to 20 -0.4498 -0.1445
#>  -0.5192  0.2152
#>  -0.7302  0.2623
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3174  0.3909 -0.3655 -0.0778 -0.2786  0.3364 -0.5267 -0.2157  0.2531
#>   0.5752  0.4952 -0.2805  0.2865 -0.5180 -0.1594 -0.6790 -0.2862  0.3801
#>   0.1058  0.3682 -0.3217  0.1333  0.1086 -0.1999 -0.8326  0.1462 -0.1293
#> 
#> Columns 10 to 18  0.2406  0.1067  0.3098 -0.1731 -0.3913 -0.1438 -0.1031 -0.2652  0.1444
#>   0.0509  0.0172  0.1141 -0.2160 -0.1310 -0.2543 -0.0740 -0.0565  0.0292
#>   0.3223  0.2098 -0.2371 -0.3352 -0.4815 -0.4917 -0.1114 -0.3495  0.3094
#> 
#> Columns 19 to 20 -0.0216  0.1308
#>  -0.0467  0.6582
#>  -0.2094 -0.0359
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>