RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6051  0.2733 -0.9327  0.3133 -0.2628 -0.5145  0.2953 -0.4168 -0.4846
#>  -0.7672  0.3155 -0.6686 -0.3860  0.4337 -0.8676 -0.6728 -0.6116  0.5152
#>   0.6194 -0.9292  0.2767 -0.0966 -0.1190 -0.8015  0.4799 -0.9039  0.6066
#> 
#> Columns 10 to 18  0.0038  0.0993  0.4432 -0.5603  0.9511  0.1858 -0.0273  0.1207 -0.9134
#>  -0.3405 -0.0659  0.5101 -0.6087  0.7879 -0.0246  0.2415  0.7955 -0.5642
#>  -0.8475 -0.7630 -0.5953  0.5947 -0.3024  0.3834  0.1749 -0.4169  0.4422
#> 
#> Columns 19 to 20  0.2583  0.2187
#>   0.9201  0.2613
#>  -0.2702  0.1289
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1977  0.6096 -0.1211  0.3493 -0.0665 -0.6414 -0.1395 -0.4877 -0.2941
#>   0.6037  0.2800 -0.3783 -0.3479 -0.0732  0.1190  0.3063 -0.5025 -0.4047
#>   0.0761 -0.2335  0.0981 -0.0230 -0.4590 -0.2536  0.3768 -0.7735  0.1791
#> 
#> Columns 10 to 18  0.1999 -0.2399 -0.4075 -0.2934  0.5134  0.8639  0.2688  0.3056  0.0471
#>  -0.1468 -0.4977 -0.3326 -0.1070  0.2081  0.6783  0.2389 -0.4611  0.0487
#>   0.1647 -0.1749 -0.4093  0.0920 -0.3793  0.3179  0.2996 -0.4892  0.1094
#> 
#> Columns 19 to 20 -0.0788  0.2738
#>   0.0118  0.5685
#>  -0.0285  0.5534
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1931  0.1575 -0.1860 -0.0126 -0.5650 -0.1720 -0.0984 -0.5861  0.4544
#>  -0.3160  0.1659  0.2491  0.0421  0.0614 -0.4471  0.0188 -0.7427  0.4645
#>  -0.2600  0.2795 -0.0379  0.1074  0.2980 -0.6355  0.0284 -0.6560 -0.1038
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1867 -0.5824 -0.0771  0.1363  0.7063 -0.2409  0.3495 -0.4202  0.3949
#>  -0.2584 -0.4480 -0.7284 -0.6890  0.1394  0.4152 -0.0894 -0.5308 -0.0826
#>   0.1964 -0.4235  0.5082  0.5065  0.8894 -0.7333 -0.0552  0.0794  0.1180
#> 
#> Columns 10 to 18 -0.4753  0.0722 -0.4773  0.2849  0.3880  0.2144  0.3709  0.4567 -0.2227
#>   0.4138  0.4196 -0.6624  0.3136  0.2996 -0.1497  0.2204 -0.0857 -0.0788
#>  -0.6789 -0.0679  0.5666 -0.4891  0.5687  0.8525 -0.4071  0.4092 -0.4778
#> 
#> Columns 19 to 20 -0.4323 -0.1077
#>  -0.3958 -0.5770
#>  -0.5681 -0.6930
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0616 -0.1011  0.1138 -0.2148 -0.1515  0.4149  0.1833 -0.6836  0.1042
#>  -0.1562 -0.3307  0.3489 -0.0771  0.0992  0.0504 -0.1690 -0.7105  0.1457
#>  -0.2855  0.3926  0.0993 -0.2521 -0.2871  0.1093  0.0776 -0.6700 -0.1047
#> 
#> Columns 10 to 18 -0.4500 -0.5270 -0.2455  0.3372  0.1079  0.3172 -0.1221 -0.1536  0.1443
#>  -0.2483 -0.5739  0.1322  0.4183 -0.3459 -0.1148 -0.1658  0.0057  0.0607
#>  -0.1318 -0.3542 -0.6355 -0.2087 -0.2172  0.2797  0.4244 -0.1239  0.1326
#> 
#> Columns 19 to 20  0.3733  0.3411
#>   0.5486  0.2728
#>  -0.0668  0.5957
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>