RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7124 -0.5769 -0.0937 -0.9371 -0.2026 -0.8579 -0.1198 -0.5769  0.4988
#>  -0.4368 -0.7549  0.1506 -0.8759 -0.6004 -0.4818 -0.0641 -0.2259 -0.7588
#>  -0.4709  0.2045  0.5078  0.7493  0.5240 -0.5179 -0.1516  0.5330  0.4116
#> 
#> Columns 10 to 18 -0.1788  0.1640 -0.8575 -0.5550  0.2347 -0.5013  0.1960  0.1809  0.5174
#>   0.1223 -0.2126 -0.4737 -0.6181 -0.7250 -0.3173 -0.6866 -0.1605  0.0931
#>  -0.2931  0.0088 -0.6096  0.2902  0.5191  0.8925  0.5663  0.0487 -0.2995
#> 
#> Columns 19 to 20  0.6231 -0.3785
#>   0.6482  0.1293
#>   0.9083  0.7630
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3261  0.4321  0.2926  0.1952  0.6910 -0.0308 -0.1527 -0.1219  0.0276
#>  -0.2778  0.2200 -0.0578  0.3731  0.6055 -0.3077  0.2825  0.2780 -0.3514
#>   0.3001  0.5124 -0.1624 -0.4582 -0.5503 -0.1471 -0.5342 -0.3692 -0.2302
#> 
#> Columns 10 to 18  0.0179 -0.5864 -0.1597  0.1636  0.5631  0.3558  0.4033  0.1563  0.1879
#>  -0.0819 -0.5117 -0.4309 -0.0556  0.3293 -0.1836  0.2065  0.3709 -0.4120
#>  -0.3266  0.3425 -0.0111 -0.6733  0.5401 -0.4644  0.5657  0.1708  0.4488
#> 
#> Columns 19 to 20  0.4291  0.6475
#>   0.1962  0.8367
#>   0.0567 -0.4624
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3434  0.2363 -0.2220 -0.4212 -0.2277 -0.4842  0.1068 -0.3341  0.3381
#>  -0.0608  0.0755 -0.4762 -0.0990  0.3358 -0.3907  0.1667  0.0314  0.4506
#>  -0.1805  0.6257  0.3788 -0.2881 -0.3470  0.1210 -0.3865 -0.2431 -0.2966
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0690  0.6076  0.8817 -0.2308  0.4395  0.6643 -0.6669  0.2920 -0.8042
#>   0.5944  0.0504 -0.4328  0.6112  0.5402  0.6357 -0.0711 -0.4338 -0.2732
#>   0.4749 -0.2465 -0.6394 -0.2745 -0.3637  0.6031  0.5570  0.0620 -0.0774
#> 
#> Columns 10 to 18 -0.8553  0.1874  0.1348 -0.3065  0.1607 -0.2206 -0.5576 -0.5778 -0.6285
#>  -0.0912  0.3287  0.0323  0.7307 -0.3819  0.1127 -0.4996 -0.1248  0.1004
#>   0.2270 -0.4018 -0.2350  0.5619  0.6276  0.7577  0.0467 -0.4188  0.4250
#> 
#> Columns 19 to 20 -0.1729 -0.6941
#>   0.3140  0.4523
#>  -0.1199  0.7334
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0178  0.1439 -0.1706 -0.3070  0.2935 -0.5559  0.1373 -0.0998  0.1843
#>   0.5780 -0.1679  0.0170 -0.1255 -0.4438 -0.4971  0.1601 -0.4211 -0.3484
#>   0.1047 -0.1314  0.2308  0.3341 -0.1008 -0.2147  0.2746 -0.5847  0.0437
#> 
#> Columns 10 to 18 -0.1346 -0.4109 -0.5663  0.1551  0.3332  0.2756 -0.1655  0.3625 -0.0560
#>  -0.1493 -0.1627 -0.1538 -0.0899  0.3096  0.2777 -0.0550  0.3323  0.2359
#>  -0.4051 -0.2462  0.2519 -0.1045 -0.1375  0.4267  0.3740  0.0911  0.1436
#> 
#> Columns 19 to 20 -0.1588  0.5528
#>  -0.3258  0.4748
#>   0.0987  0.6983
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>