RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3451  0.2345  0.4944 -0.6933  0.2857  0.5165 -0.3727 -0.1310 -0.3172
#>   0.3932 -0.2771  0.3422 -0.7802  0.2011  0.7937  0.1520 -0.5842 -0.1160
#>   0.4775  0.3998  0.0734 -0.3068 -0.2362 -0.0990  0.2741  0.0320  0.0844
#> 
#> Columns 10 to 18 -0.6241  0.0625 -0.1393  0.0498 -0.0069  0.2726 -0.3930 -0.0747 -0.8710
#>   0.0904 -0.5788  0.1254 -0.4327 -0.1644 -0.0561  0.6780 -0.8153  0.4367
#>   0.5787 -0.2841  0.5185 -0.5281 -0.0972  0.3106  0.4641  0.2365  0.1235
#> 
#> Columns 19 to 20  0.7990  0.0821
#>   0.3709 -0.1383
#>   0.1495 -0.6741
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0700  0.3400 -0.5457  0.0073 -0.2481 -0.0442 -0.5436  0.1110 -0.1892
#>   0.3376 -0.5519 -0.0296  0.4520  0.2532 -0.1793  0.0483 -0.2597 -0.2996
#>   0.7670  0.0579 -0.0168 -0.0649 -0.1959  0.1609  0.0227 -0.0045 -0.2881
#> 
#> Columns 10 to 18  0.0994 -0.3818 -0.2365  0.0989 -0.2684  0.1773  0.1915  0.0980  0.4837
#>  -0.2287 -0.1285 -0.1176  0.1659 -0.7948  0.1957  0.0563  0.0492 -0.0741
#>  -0.1987 -0.0886  0.1744  0.2065 -0.3733  0.1888 -0.1908  0.0257  0.1156
#> 
#> Columns 19 to 20  0.0045  0.2527
#>  -0.0125  0.5340
#>   0.3651  0.3396
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2329 -0.2660  0.1176 -0.0161  0.2602  0.3492  0.3143 -0.3373 -0.3953
#>  -0.0128 -0.0223 -0.1014  0.1402  0.2991 -0.1707 -0.1941  0.1245 -0.0022
#>   0.0867 -0.4219  0.1188  0.2184  0.2370  0.6044  0.0306 -0.3656 -0.4988
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3436  0.3669 -0.4164 -0.4227 -0.2547 -0.0650 -0.5899 -0.6061 -0.6110
#>   0.0644  0.4934  0.0861 -0.2570  0.2057 -0.6734  0.2659  0.2964 -0.3538
#>  -0.1421  0.2590  0.4169 -0.1774  0.6307 -0.6851 -0.1629  0.6514  0.3017
#> 
#> Columns 10 to 18  0.3652  0.3344 -0.3249  0.3348  0.2724  0.5803  0.3601 -0.5374  0.0966
#>  -0.6223 -0.5815  0.3590 -0.3653  0.1564 -0.2248  0.0102  0.2215  0.6308
#>  -0.5388 -0.2627  0.0930 -0.2097  0.3271 -0.2368 -0.1716 -0.2770  0.3264
#> 
#> Columns 19 to 20  0.5032 -0.6907
#>   0.3113 -0.1808
#>   0.3272  0.4528
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2931  0.0703  0.1442  0.5077  0.3567  0.3249  0.2448 -0.4420 -0.2527
#>   0.1061 -0.5408  0.3698  0.1516  0.2683  0.0554 -0.1988 -0.4177 -0.0068
#>   0.5232 -0.6217  0.0735 -0.1668  0.4383  0.2173  0.2211 -0.2711 -0.1731
#> 
#> Columns 10 to 18  0.0470 -0.1228  0.0013  0.4056 -0.0659  0.3843 -0.1313  0.3755  0.5870
#>  -0.2531  0.0331  0.2341  0.3623 -0.1095  0.0097 -0.0680 -0.0417  0.3302
#>  -0.0808 -0.1988  0.4545  0.0578 -0.1974 -0.0916 -0.0118 -0.2482  0.0793
#> 
#> Columns 19 to 20 -0.0225  0.3477
#>   0.1973  0.1594
#>  -0.1389 -0.1311
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>