RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3473  0.0645 -0.6913  0.7706  0.4696  0.0443 -0.2276  0.6972  0.3707
#>   0.0785 -0.7003  0.6803  0.0848 -0.7474  0.4358  0.5591 -0.1405  0.6656
#>  -0.6552  0.1263  0.6147  0.2806 -0.3971  0.8095 -0.4225 -0.5992 -0.7672
#> 
#> Columns 10 to 18 -0.4756 -0.5694 -0.0912  0.2056  0.3408 -0.4738 -0.5243 -0.4441  0.5509
#>   0.8313 -0.4586  0.5777  0.4365 -0.2952 -0.2284  0.3596 -0.2451  0.5312
#>   0.2002  0.2653 -0.2830 -0.1385 -0.7124 -0.7995 -0.3605 -0.1807  0.1066
#> 
#> Columns 19 to 20 -0.0807 -0.0414
#>  -0.1238  0.3145
#>  -0.8626 -0.3058
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2069 -0.3193 -0.2262  0.5084 -0.1665  0.6866  0.0375  0.0771  0.1997
#>  -0.5121  0.1388 -0.1477  0.2488 -0.2942  0.0995 -0.6009 -0.5031  0.8475
#>  -0.2468  0.3657  0.6644  0.1100 -0.5483  0.5380 -0.4950  0.3526 -0.2971
#> 
#> Columns 10 to 18  0.1145  0.4600  0.0690  0.1547 -0.2216  0.0430  0.0094  0.1782  0.7117
#>   0.0659  0.3769 -0.0393  0.2758  0.0429  0.2604  0.3476  0.0831  0.3881
#>  -0.2039  0.7030 -0.5780  0.0998  0.7617  0.0167 -0.3354  0.5912 -0.5697
#> 
#> Columns 19 to 20  0.3824 -0.3004
#>  -0.2231 -0.5598
#>  -0.5061 -0.2006
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2664  0.3260  0.4528  0.1599 -0.0400  0.5816 -0.5129  0.3128 -0.0733
#>   0.2434 -0.0820 -0.0604  0.1266 -0.7130  0.6452 -0.2555 -0.2547  0.1067
#>   0.4770  0.3927  0.2388  0.3461 -0.4328  0.8300  0.1653  0.6843  0.1427
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6812  0.8076  0.4520  0.5761  0.0792  0.4458 -0.2395 -0.0455  0.6702
#>   0.3251  0.3167 -0.1210  0.1169 -0.0257  0.3517  0.4710 -0.2241 -0.3769
#>   0.4569  0.4987  0.4499  0.1488 -0.1017  0.3628 -0.4712  0.4140  0.3433
#> 
#> Columns 10 to 18 -0.5447  0.4592 -0.4533  0.0062  0.3162  0.0989  0.7119  0.2636 -0.6693
#>   0.7017 -0.1009  0.5783  0.0527 -0.1558  0.4633  0.5328 -0.1632  0.6346
#>  -0.2624  0.4712  0.4499 -0.4116  0.0918  0.2138  0.1221 -0.4824  0.1531
#> 
#> Columns 19 to 20 -0.7134  0.1647
#>   0.5250  0.5661
#>  -0.2030 -0.4228
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5674  0.5228  0.5723  0.3916  0.1887  0.6347 -0.3931  0.5172 -0.3695
#>   0.2228  0.5574  0.3191  0.1812 -0.3337  0.6186 -0.0536  0.4280 -0.0207
#>  -0.0607  0.6527  0.3617  0.5018  0.1195  0.5481 -0.5363  0.5605 -0.1918
#> 
#> Columns 10 to 18  0.0558  0.4839 -0.1642 -0.2997 -0.2778 -0.2414 -0.7258 -0.6978 -0.1861
#>   0.2660  0.5543 -0.2823 -0.1327 -0.1686  0.2029 -0.2665  0.1017  0.0003
#>   0.1178  0.3950 -0.4120 -0.4162 -0.2922  0.0714 -0.6538 -0.2314 -0.3747
#> 
#> Columns 19 to 20  0.2769  0.7351
#>   0.3394  0.2764
#>   0.4643  0.1591
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>