RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2200  0.6174 -0.7095 -0.1982 -0.4190  0.4147 -0.4139 -0.4970 -0.4431
#>   0.4293 -0.5940 -0.7462 -0.1882  0.5914  0.0296 -0.7651  0.1193 -0.5813
#>  -0.5851  0.4078 -0.4438  0.6617 -0.1977 -0.0398  0.2591 -0.3858 -0.4935
#> 
#> Columns 10 to 18  0.3641 -0.7327 -0.5284 -0.3542  0.1377  0.5778  0.5524  0.4849 -0.3649
#>   0.7996  0.4798 -0.7522 -0.9240 -0.2330  0.7094  0.2865  0.5701  0.6830
#>  -0.0484 -0.2661  0.2667  0.5731 -0.7658 -0.0791 -0.6810 -0.8880 -0.1887
#> 
#> Columns 19 to 20  0.4744 -0.5380
#>  -0.2930  0.7232
#>   0.1526 -0.2759
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1365 -0.1297  0.5271  0.3920  0.4387  0.6315 -0.1373  0.0208  0.0824
#>  -0.3814  0.3863 -0.4246  0.4609 -0.5539  0.0561 -0.0691  0.1300 -0.2514
#>  -0.1896  0.0733 -0.4687 -0.5223  0.2254  0.3001 -0.1749 -0.0449  0.0027
#> 
#> Columns 10 to 18 -0.0654 -0.5991 -0.3032  0.2870 -0.6127 -0.0407 -0.2037 -0.1006  0.5711
#>   0.0025 -0.2230 -0.5472  0.1735  0.0283  0.8414 -0.3431 -0.0363  0.0247
#>   0.0349  0.1877 -0.0697 -0.1962 -0.3771  0.3149  0.1808  0.1338  0.4596
#> 
#> Columns 19 to 20  0.4803 -0.6347
#>   0.0634 -0.4384
#>   0.3982  0.3824
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0758 -0.3269 -0.3704 -0.0099  0.2585  0.2502  0.1693 -0.0634  0.5017
#>  -0.0931 -0.0526  0.0502  0.2816  0.2545  0.4446  0.1862 -0.0807  0.4388
#>   0.1911 -0.3534  0.0188  0.1919  0.1726  0.0744 -0.3555  0.3089 -0.4892
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8039 -0.1286  0.2395 -0.7113  0.3557 -0.0213 -0.6344 -0.4922 -0.1586
#>  -0.2981  0.6351 -0.5277  0.1553 -0.0374 -0.4946  0.4619 -0.3140 -0.3967
#>  -0.7091  0.5527  0.0290 -0.5990 -0.2408 -0.3499  0.1798 -0.2137  0.1263
#> 
#> Columns 10 to 18  0.0136 -0.0883 -0.1030  0.4556 -0.5730  0.2424 -0.7922  0.1597 -0.6178
#>   0.1547 -0.1342 -0.4015  0.1212 -0.5870  0.6233  0.3580  0.5299  0.4582
#>  -0.1885  0.2651 -0.3456 -0.4260 -0.3178 -0.3551  0.7743 -0.3123  0.6683
#> 
#> Columns 19 to 20  0.4308 -0.7076
#>  -0.0451 -0.4776
#>   0.2783 -0.2366
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0191 -0.0115  0.2195  0.2979  0.4005  0.3026  0.2999 -0.0180  0.1301
#>  -0.2840  0.5305 -0.1925  0.0474  0.0531  0.3352 -0.1754 -0.2708  0.3194
#>  -0.3950  0.0448 -0.6871  0.0052 -0.1511  0.4766  0.0935 -0.0410 -0.3333
#> 
#> Columns 10 to 18  0.4786 -0.1937 -0.5120 -0.0116 -0.6007  0.4183 -0.0230  0.0589  0.5935
#>  -0.0309  0.1592 -0.4947 -0.1805 -0.2594  0.6359 -0.1660  0.4090  0.6475
#>   0.2177 -0.0005 -0.4117 -0.1844 -0.4991  0.3532 -0.4543 -0.3227  0.3157
#> 
#> Columns 19 to 20  0.2165  0.1225
#>   0.3635  0.3618
#>   0.6118 -0.4524
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>