RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2964 -0.8875 -0.3283  0.0381  0.6832  0.6649  0.2109 -0.1637 -0.6178
#>   0.7073  0.2763 -0.7768 -0.7380  0.6423 -0.4511 -0.2317 -0.1600 -0.1369
#>  -0.9051 -0.6535  0.0785  0.3973 -0.4242  0.7159 -0.9393  0.5245 -0.7821
#> 
#> Columns 10 to 18  0.6597 -0.7891  0.7124  0.1646 -0.6256 -0.8101 -0.1935 -0.0096  0.1426
#>   0.4229 -0.3695  0.3005 -0.5462 -0.5048  0.7871  0.4256  0.3938 -0.4089
#>   0.0440 -0.1103 -0.0528  0.8662 -0.9654 -0.5071  0.2673  0.3370  0.6727
#> 
#> Columns 19 to 20  0.6641 -0.6353
#>  -0.8890 -0.8218
#>   0.8246 -0.8459
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3764 -0.0782 -0.1045 -0.4747  0.6592 -0.0718  0.0496  0.2803  0.0974
#>  -0.6397  0.1472  0.0596 -0.5710  0.5139  0.7276 -0.3212  0.0788 -0.4141
#>  -0.0259 -0.1882 -0.6502 -0.7077  0.5420  0.0058  0.2819  0.2340  0.3022
#> 
#> Columns 10 to 18  0.4473 -0.1515  0.0729  0.2875 -0.6578 -0.3373 -0.3736  0.0684  0.0843
#>   0.2342 -0.1134  0.6351  0.2637 -0.7459 -0.4067  0.4512 -0.1963  0.6436
#>  -0.0017  0.0427  0.2710  0.7673 -0.5168 -0.3391 -0.0363  0.4543 -0.1710
#> 
#> Columns 19 to 20 -0.2300 -0.2586
#>   0.8377  0.1781
#>  -0.5305  0.3540
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.7229  0.3332 -0.2824 -0.6979  0.7199  0.7549  0.0814  0.5392 -0.2033
#>   0.3343 -0.4322 -0.0851 -0.4200  0.6084  0.1985 -0.3856 -0.1868  0.5010
#>   0.1522  0.0771  0.3092 -0.4903  0.4693 -0.2381 -0.2020 -0.2228  0.1764
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3938  0.8699 -0.3752 -0.0434 -0.7706  0.3481 -0.0403  0.1689 -0.5824
#>   0.3875 -0.2857 -0.5427 -0.7570  0.6303  0.3374  0.1189  0.3561 -0.8014
#>   0.5934 -0.2053 -0.8217  0.5902  0.2898  0.4557  0.5175  0.1345 -0.2033
#> 
#> Columns 10 to 18  0.3659  0.0506 -0.0254 -0.4320  0.2720  0.0354 -0.6760  0.3729  0.4368
#>   0.4592 -0.1529 -0.3854 -0.6207 -0.7503  0.7095  0.5663 -0.2735  0.6026
#>   0.7997 -0.2993 -0.1026  0.0143  0.8048  0.2368 -0.1132 -0.0111  0.4844
#> 
#> Columns 19 to 20 -0.8075  0.6143
#>  -0.0560  0.1078
#>   0.4325 -0.6315
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3419  0.2322  0.2409 -0.2500  0.2926 -0.3437  0.0773  0.4738 -0.1474
#>  -0.0149 -0.1826 -0.0637 -0.6387  0.7945  0.7248  0.1647  0.0871  0.4433
#>   0.8093  0.0018 -0.2561 -0.0444 -0.2141  0.2465 -0.3772  0.0808  0.1027
#> 
#> Columns 10 to 18  0.3663 -0.1952  0.0388  0.2141  0.1329 -0.5094 -0.2726 -0.0951  0.1149
#>   0.7054 -0.3889  0.2244  0.4741 -0.6868 -0.6121 -0.1573 -0.3215  0.0849
#>   0.0430 -0.4997 -0.1585  0.3099 -0.6883 -0.4405 -0.2622  0.1991 -0.2591
#> 
#> Columns 19 to 20 -0.2912 -0.3348
#>   0.7423  0.3806
#>  -0.2734  0.2137
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>