RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9071  0.6026  0.2784  0.2329 -0.1999  0.9813  0.0457 -0.7215 -0.6252
#>   0.6313 -0.5922 -0.8362  0.6870 -0.1103 -0.3664 -0.3085 -0.0700 -0.0155
#>  -0.7347  0.1716  0.4776  0.2904  0.4526 -0.2547 -0.2326  0.3158 -0.0614
#> 
#> Columns 10 to 18 -0.3365  0.3862  0.4831  0.5914  0.7074  0.5403 -0.1762 -0.6750  0.5775
#>   0.3046 -0.2737 -0.7824  0.1625  0.3736 -0.5457 -0.1051  0.6444  0.8264
#>  -0.6215 -0.7849 -0.6882 -0.3372  0.9066  0.2052 -0.5423 -0.1188  0.5109
#> 
#> Columns 19 to 20 -0.8027  0.4664
#>  -0.3107  0.5803
#>   0.7725 -0.6981
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3827 -0.0938 -0.5338  0.2050 -0.1668  0.0938  0.1848 -0.1909 -0.3386
#>   0.3225  0.1368 -0.4954 -0.3504  0.3935  0.5243  0.0637  0.1644  0.4680
#>   0.8092 -0.1987 -0.0514 -0.5483  0.1259  0.2827  0.3969 -0.1690  0.4819
#> 
#> Columns 10 to 18  0.1151  0.5686  0.2264  0.3457  0.0086 -0.1598  0.0065  0.4950 -0.5230
#>   0.5326 -0.0010  0.5358  0.3317 -0.2354  0.3700 -0.1223 -0.3146 -0.1434
#>   0.7379  0.1318 -0.4639  0.0503 -0.1251 -0.4590  0.1181  0.7005  0.8359
#> 
#> Columns 19 to 20 -0.1381  0.1416
#>  -0.6409  0.0690
#>   0.1733 -0.1698
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1185 -0.0879 -0.3455  0.1295 -0.5333  0.6159 -0.1365 -0.1794  0.0070
#>  -0.1416 -0.0354  0.4271  0.0317  0.1907  0.5272  0.0145 -0.2417  0.1263
#>   0.1361  0.4548 -0.3761 -0.3095  0.0216  0.2112  0.5438 -0.0710 -0.0933
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6488  0.5336  0.0712  0.1702 -0.6854 -0.8312 -0.5376  0.5964 -0.3633
#>   0.7641 -0.7651 -0.6870 -0.4175  0.6951  0.0032 -0.6519  0.3299  0.2800
#>  -0.3913 -0.0529  0.0240 -0.1338 -0.4176  0.0433 -0.6288 -0.3333  0.4702
#> 
#> Columns 10 to 18 -0.6289  0.1080  0.2179  0.5951  0.6852  0.3981 -0.5539  0.6082  0.2915
#>   0.5204  0.7136 -0.5672 -0.5868  0.7523  0.8648  0.4136 -0.3862  0.4158
#>  -0.0873  0.3692  0.4111 -0.3776  0.1167  0.4337 -0.0588 -0.5860 -0.1372
#> 
#> Columns 19 to 20 -0.3492  0.7219
#>   0.6891  0.4150
#>   0.3565  0.5200
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2449 -0.4029 -0.2510 -0.1283 -0.3726  0.8232 -0.5408 -0.4185  0.0542
#>  -0.0386  0.2180 -0.4742  0.0782 -0.5324  0.1967 -0.0280 -0.1418  0.2938
#>   0.4654 -0.1158 -0.2913 -0.2953 -0.1208  0.0642 -0.1346  0.0190  0.3792
#> 
#> Columns 10 to 18 -0.2274  0.3982 -0.4717  0.0252  0.0924 -0.2642  0.1850 -0.2209 -0.5146
#>   0.4729  0.2676  0.2155  0.7388  0.4984  0.7189  0.1249  0.2095  0.2277
#>   0.4859  0.3237 -0.0605  0.6227  0.1034  0.1473 -0.1654 -0.1194  0.4162
#> 
#> Columns 19 to 20 -0.2378  0.7088
#>  -0.1920  0.1424
#>  -0.3387  0.0108
#> [ CPUFloatType{2,3,20} ]
#>