RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6648 -0.2898  0.6735  0.1494  0.2316  0.3076  0.5917  0.8948 -0.0080
#>   0.0691  0.0478  0.6836 -0.0613 -0.0464 -0.5567 -0.1052 -0.0168 -0.2447
#>  -0.2854 -0.8283  0.9355  0.8190  0.8808  0.6587  0.3678 -0.0068 -0.2660
#> 
#> Columns 10 to 18 -0.1800  0.0755  0.3651 -0.1305  0.6327  0.8958  0.7508 -0.2988 -0.7818
#>  -0.6476  0.1365  0.2477  0.8251 -0.0551  0.5067 -0.1177 -0.3102 -0.7830
#>   0.7141  0.5556 -0.0997 -0.2609 -0.0589 -0.1500 -0.7907 -0.8287 -0.2162
#> 
#> Columns 19 to 20 -0.5849  0.1781
#>   0.3695 -0.6049
#>   0.3632 -0.3061
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1166  0.4613 -0.2282  0.4001  0.3784  0.3393 -0.5880  0.7320  0.1827
#>   0.1831 -0.3334  0.7172  0.3172 -0.1179  0.2413 -0.0592  0.2887 -0.0351
#>   0.4715  0.0971  0.6892 -0.2906  0.1570 -0.4466  0.3637  0.0085 -0.2936
#> 
#> Columns 10 to 18 -0.4315 -0.5048 -0.3909 -0.0463  0.1027 -0.3535  0.5609 -0.6806 -0.1296
#>   0.1102  0.2175  0.2522 -0.4548  0.3473  0.1910 -0.0964  0.1355  0.0753
#>  -0.7245 -0.4345 -0.5149 -0.2411  0.3834 -0.3993  0.0604 -0.3930 -0.0956
#> 
#> Columns 19 to 20  0.4060  0.3448
#>   0.4697  0.0361
#>   0.3125  0.3823
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1176 -0.0474  0.1910  0.3064  0.3915  0.0483  0.5789 -0.1970 -0.5220
#>   0.1938 -0.1058  0.6359  0.1200  0.2611  0.2618  0.5948  0.1497 -0.4114
#>   0.3599 -0.2932  0.5928  0.5984 -0.0718  0.2566  0.4535  0.0231 -0.0056
#> 
#> Columns 10 to 18 -0.2994  0.0247 -0.6300  0.3924  0.1397  0.0807  0.0767 -0.3727  0.2083
#>  -0.2587 -0.1480 -0.5032 -0.2043  0.0792  0.0900 -0.0345 -0.1460 -0.1574
#>  -0.3896 -0.4305  0.1377 -0.0460 -0.0255 -0.0725 -0.1173 -0.0959 -0.3475
#> 
#> Columns 19 to 20  0.5492 -0.1625
#>   0.3618 -0.0141
#>   0.5141 -0.5139
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.4038 -0.0353  0.3907  0.1740 -0.0296  0.3622  0.3162 -0.1254  0.2241
#>   0.3073  0.0434  0.2360  0.2481  0.2841  0.3761 -0.1387  0.4988 -0.1150
#>   0.5232  0.0495  0.2988  0.4553 -0.1994  0.4292  0.5748  0.0733 -0.2295
#> 
#> Columns 10 to 18 -0.4251 -0.1670  0.0784 -0.5104 -0.0614 -0.3193 -0.3216 -0.0841 -0.1548
#>  -0.6631 -0.6725 -0.4005 -0.3659  0.1843 -0.2037  0.4066 -0.1950 -0.4962
#>  -0.2504 -0.3201 -0.2237 -0.5795  0.1038 -0.1810 -0.1743  0.2799 -0.1447
#> 
#> Columns 19 to 20  0.1852  0.1483
#>   0.4195 -0.2442
#>   0.5716  0.0406
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.4847  0.2624  0.3819  0.0632 -0.1442  0.2764  0.1955  0.7238  0.0144
#>   0.5483  0.1424  0.1014  0.2099  0.0570  0.2690  0.8126 -0.3011 -0.4138
#>   0.6343  0.3210  0.2910  0.0897 -0.0563  0.4871  0.4288  0.5725 -0.4767
#> 
#> Columns 10 to 18 -0.3002 -0.1977 -0.3173 -0.3104  0.1587 -0.2143  0.1587 -0.2831 -0.5840
#>  -0.1859 -0.0587 -0.2563  0.0326 -0.2769  0.1994  0.3180 -0.1369 -0.5265
#>  -0.1895 -0.3455 -0.5426 -0.3082  0.0033  0.0963  0.0684 -0.3901 -0.4815
#> 
#> Columns 19 to 20  0.6020 -0.0752
#>   0.5717 -0.1635
#>   0.5000 -0.1524
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1193 -0.1953 -0.6291  0.7009  0.4436  0.2538 -0.1303  0.3212 -0.6101
#>  -0.1381 -0.2035  0.2790  0.3287 -0.2832 -0.4964  0.5168 -0.4523 -0.4460
#>  -0.4664 -0.4467  0.1892  0.7904 -0.1746  0.2710 -0.5299 -0.2420  0.2831
#> 
#> Columns 10 to 18 -0.6124 -0.5033  0.4940 -0.1047 -0.0789 -0.5378 -0.3559  0.6869 -0.3350
#>   0.4135  0.3285 -0.4135  0.4767 -0.0003 -0.3849 -0.5481  0.8301 -0.4292
#>   0.2316 -0.4385 -0.2559  0.5747 -0.6873 -0.5083  0.0487  0.1423 -0.5018
#> 
#> Columns 19 to 20 -0.4390 -0.5964
#>  -0.5941 -0.0742
#>  -0.1714  0.0537
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4847  0.2624  0.3819  0.0632 -0.1442  0.2764  0.1955  0.7238  0.0144
#>   0.5483  0.1424  0.1014  0.2099  0.0570  0.2690  0.8126 -0.3011 -0.4138
#>   0.6343  0.3210  0.2910  0.0897 -0.0563  0.4871  0.4288  0.5725 -0.4767
#> 
#> Columns 10 to 18 -0.3002 -0.1977 -0.3173 -0.3104  0.1587 -0.2143  0.1587 -0.2831 -0.5840
#>  -0.1859 -0.0587 -0.2563  0.0326 -0.2769  0.1994  0.3180 -0.1369 -0.5265
#>  -0.1895 -0.3455 -0.5426 -0.3082  0.0033  0.0963  0.0684 -0.3901 -0.4815
#> 
#> Columns 19 to 20  0.6020 -0.0752
#>   0.5717 -0.1635
#>   0.5000 -0.1524
#> [ CPUFloatType{2,3,20} ]
#>