RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7774 -0.6234 -0.8830 -0.7408 -0.9053 -0.4849  0.7110  0.1017 -0.8042
#>   0.3790 -0.4079 -0.6935 -0.4647 -0.8907  0.1691  0.5714 -0.2475 -0.3098
#>   0.6078  0.3002 -0.4131 -0.7701 -0.0969 -0.8560  0.3570  0.9221 -0.7040
#> 
#> Columns 10 to 18 -0.4688  0.3419  0.3944 -0.4749  0.0016 -0.3356  0.0026  0.5515 -0.7168
#>  -0.5599  0.2639 -0.3005  0.3181 -0.6558  0.0692  0.0916  0.7403 -0.0557
#>   0.4088  0.1277  0.6169 -0.9403 -0.8671  0.0468 -0.9423 -0.3169 -0.8288
#> 
#> Columns 19 to 20 -0.3363  0.2158
#>   0.0448  0.8534
#>   0.1964  0.0062
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0494  0.4396  0.1548  0.2176 -0.8864  0.0396 -0.2354  0.0140  0.0437
#>   0.1903 -0.0491 -0.2784  0.0584 -0.6517  0.2778 -0.5101  0.2116  0.1426
#>   0.6061 -0.5965 -0.4223 -0.2582 -0.7906 -0.6082  0.7656  0.7525 -0.6169
#> 
#> Columns 10 to 18  0.2083  0.5261  0.2832  0.2381  0.3918  0.2770 -0.1245  0.6965 -0.0801
#>   0.3443 -0.1453  0.2778  0.0713 -0.5249 -0.5967  0.1619 -0.0270 -0.3767
#>  -0.2295  0.2619  0.3495 -0.7670 -0.3302  0.4069  0.1723  0.5774 -0.3244
#> 
#> Columns 19 to 20 -0.4662  0.1287
#>  -0.1840  0.2499
#>  -0.0193 -0.1495
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0713 -0.3224 -0.2667 -0.1161 -0.5729 -0.2295  0.1353 -0.3279 -0.0454
#>   0.0786 -0.4962 -0.5313 -0.0667 -0.7138 -0.4156  0.5122  0.3516 -0.4299
#>   0.6681 -0.0157  0.2072 -0.4773 -0.2308 -0.0071 -0.1014  0.1432 -0.6207
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8994 -0.6361 -0.0057 -0.3380  0.3734 -0.1368  0.6299 -0.2172 -0.0850
#>   0.2057 -0.6893 -0.0901  0.0984 -0.8702 -0.8000 -0.5290 -0.0001 -0.3401
#>   0.4903 -0.8333 -0.1798 -0.7048 -0.0940 -0.2593  0.0093  0.3902 -0.0050
#> 
#> Columns 10 to 18  0.5004  0.5075 -0.2892  0.1763 -0.3038  0.5167  0.4361  0.3702 -0.6436
#>   0.8097  0.6799 -0.2853  0.1311 -0.2006 -0.3887 -0.5793 -0.4486  0.4345
#>   0.3850  0.5062 -0.3125 -0.0722 -0.0633  0.2692 -0.5121  0.1054 -0.0422
#> 
#> Columns 19 to 20 -0.6315  0.6826
#>  -0.0619 -0.0197
#>  -0.2637 -0.0028
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2918 -0.0791 -0.5381  0.2192 -0.8854 -0.1867  0.4160 -0.1212  0.2476
#>   0.7208 -0.3737 -0.0252 -0.4752 -0.3333 -0.1712 -0.2023 -0.0451 -0.2899
#>   0.1455 -0.0977 -0.1586 -0.4925 -0.8584 -0.2632 -0.1045 -0.1051 -0.2179
#> 
#> Columns 10 to 18  0.0978  0.0873  0.1919  0.2715 -0.1293  0.0104 -0.0613  0.4781 -0.3247
#>  -0.2941  0.2455  0.3808 -0.2386 -0.3712 -0.4025 -0.3091  0.6441 -0.2135
#>  -0.0576  0.5108  0.0393 -0.2449 -0.2027 -0.2817 -0.2418  0.6616 -0.4132
#> 
#> Columns 19 to 20 -0.5567  0.2199
#>  -0.3601  0.0707
#>  -0.2171  0.3620
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>