RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5164  0.1072 -0.3735 -0.1782  0.0033  0.0900 -0.0085 -0.7453 -0.2728
#>   0.2346 -0.4938 -0.7227 -0.0347 -0.1173 -0.5629  0.0325  0.1953 -0.3818
#>   0.3445  0.0109 -0.5327 -0.9006  0.3409 -0.4495 -0.0039 -0.3461  0.4366
#> 
#> Columns 10 to 18  0.1613  0.0090 -0.2281 -0.7952 -0.5376  0.8493 -0.4794 -0.7639  0.4143
#>  -0.7145  0.7277  0.7018 -0.5962 -0.6033  0.8441 -0.8632 -0.4925 -0.2912
#>  -0.2243 -0.2468  0.8081  0.2865  0.7705 -0.1247 -0.0947 -0.1771 -0.5870
#> 
#> Columns 19 to 20  0.2167 -0.2074
#>   0.1489 -0.4514
#>  -0.8309 -0.0312
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6720  0.1926 -0.5369  0.0764  0.0329  0.3641 -0.6503  0.1321  0.2158
#>   0.4333  0.2338 -0.7204 -0.6094  0.1274  0.6499 -0.1690  0.0565 -0.0274
#>  -0.5827  0.2096 -0.2901  0.1024 -0.0850  0.1409 -0.3189 -0.0123  0.0931
#> 
#> Columns 10 to 18  0.1884  0.2913  0.5208 -0.1150 -0.0812 -0.3017  0.2312 -0.4228  0.0302
#>   0.1211  0.4680  0.3229 -0.3109 -0.2019 -0.1652  0.5100 -0.2260 -0.3946
#>   0.1868 -0.3522  0.5578 -0.1070  0.0846  0.0163  0.5930 -0.5153 -0.3147
#> 
#> Columns 19 to 20  0.4314 -0.4810
#>   0.7040 -0.1282
#>   0.4207 -0.1541
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0091  0.0560 -0.2218  0.1597 -0.4826  0.6820 -0.0119 -0.1092 -0.0258
#>  -0.0497  0.1256 -0.1314  0.1149 -0.1167  0.2553 -0.2268 -0.3093 -0.0011
#>   0.0298 -0.0974  0.0648 -0.0081 -0.2571  0.3208  0.3566 -0.6168  0.0088
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4674  0.2500  0.5661 -0.4105 -0.2009 -0.4345  0.0121  0.0425 -0.0888
#>   0.1943 -0.0731 -0.3077 -0.1441  0.3171  0.6993 -0.1129 -0.6526 -0.0745
#>   0.4875 -0.8131 -0.0463 -0.6651  0.3853  0.1385 -0.5093  0.1841  0.7123
#> 
#> Columns 10 to 18 -0.1416 -0.3559  0.2791  0.4046  0.4560  0.2148  0.3225  0.2589 -0.1544
#>   0.3533 -0.1659 -0.7258 -0.7939 -0.3316  0.5309 -0.0862 -0.3902 -0.4660
#>  -0.8363 -0.6491 -0.1096 -0.7089 -0.5727  0.1278  0.2023 -0.7150 -0.9045
#> 
#> Columns 19 to 20 -0.1224  0.4687
#>  -0.4487  0.2957
#>  -0.6438  0.3734
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1542  0.0278 -0.0891  0.3418 -0.1548  0.4339 -0.0378 -0.2934  0.0874
#>  -0.1882 -0.1120 -0.2431 -0.6459  0.0157 -0.1544  0.0216 -0.4109 -0.0895
#>  -0.1397 -0.0652 -0.0941 -0.3703  0.0474  0.2752  0.3527 -0.3938 -0.2476
#> 
#> Columns 10 to 18  0.2716 -0.1480  0.5293 -0.2754  0.0267  0.0975  0.2391 -0.3512 -0.0688
#>   0.2010 -0.2908  0.4191 -0.0862  0.5223 -0.1525  0.3992  0.1586 -0.0170
#>  -0.4777 -0.4378  0.7938 -0.5386 -0.0072  0.4824  0.2444 -0.4891 -0.5172
#> 
#> Columns 19 to 20  0.0544 -0.2636
#>   0.2362 -0.1196
#>   0.0023 -0.3347
#> [ CPUFloatType{2,3,20} ]
#>