RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2875  0.5386  0.6612  0.8247 -0.5576 -0.8259 -0.8492  0.0043  0.1679
#>  -0.0813  0.4930  0.3183  0.8339  0.3105 -0.4055 -0.4478  0.1338 -0.3739
#>  -0.8259 -0.5474  0.5463  0.1927 -0.5990 -0.2283 -0.0073  0.3738 -0.0790
#> 
#> Columns 10 to 18 -0.4154  0.0218 -0.3412 -0.1814 -0.0708  0.2074 -0.3215 -0.6141  0.2424
#>  -0.7417 -0.5454 -0.1057 -0.3543  0.3671 -0.8928  0.4552 -0.3351 -0.5320
#>  -0.4353 -0.3236 -0.1296  0.4387  0.6996 -0.4791  0.7426 -0.7110  0.1493
#> 
#> Columns 19 to 20  0.1815  0.9108
#>  -0.8479  0.6570
#>   0.0475  0.4647
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0097 -0.1472 -0.5631  0.3952  0.1738  0.0197 -0.1021  0.7411  0.0734
#>   0.0684 -0.5581 -0.0148 -0.3252 -0.1823 -0.1168 -0.2592  0.1044  0.5197
#>  -0.3013  0.2926 -0.0531 -0.3352  0.2976 -0.4724 -0.0120  0.4170 -0.6220
#> 
#> Columns 10 to 18 -0.2013  0.3491  0.1355  0.1457 -0.8098 -0.0189 -0.2025 -0.1177 -0.4962
#>   0.2708 -0.4557  0.6143 -0.0983 -0.2720  0.4461 -0.2164 -0.1982 -0.3076
#>  -0.0226 -0.3823  0.3321  0.0024  0.2879  0.0160 -0.1805  0.3002 -0.5294
#> 
#> Columns 19 to 20 -0.1807  0.2177
#>  -0.5237  0.0443
#>  -0.6519  0.4673
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3712  0.5092 -0.2142  0.2444  0.1571  0.1064 -0.1534  0.5025 -0.5172
#>  -0.2365  0.5946 -0.2797 -0.2843 -0.2831 -0.0998 -0.4099  0.3574 -0.5268
#>   0.0283  0.4500 -0.2720 -0.5813 -0.3816 -0.2166 -0.1783 -0.2291 -0.0597
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4278  0.0476  0.3591  0.2205  0.6557 -0.6507 -0.2170 -0.2363 -0.1365
#>  -0.2925 -0.3541  0.6294 -0.1601  0.2919 -0.6670 -0.1975  0.3875  0.4244
#>   0.1522  0.0225  0.1921  0.0036  0.4684 -0.7364  0.0922  0.4701  0.2984
#> 
#> Columns 10 to 18  0.3270  0.2912 -0.0108 -0.2029  0.0664 -0.0265 -0.2948  0.5492  0.3698
#>   0.2178  0.8429  0.1383  0.0778 -0.1529  0.4630 -0.1249  0.5984  0.0520
#>  -0.1362  0.3643  0.5435 -0.2050 -0.0150  0.4774 -0.2673  0.4156 -0.3652
#> 
#> Columns 19 to 20  0.3396  0.5228
#>   0.2751 -0.2132
#>  -0.0254  0.2119
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1216  0.1382 -0.0372  0.1577 -0.4093 -0.2429 -0.2843  0.2333 -0.2089
#>   0.0768  0.3399  0.0328 -0.4037 -0.3622  0.2039 -0.3244  0.5228 -0.3147
#>   0.2595  0.3451 -0.2331 -0.3008 -0.5881 -0.1224 -0.1419  0.1908 -0.1629
#> 
#> Columns 10 to 18 -0.2024  0.2460  0.3137  0.3573  0.2409 -0.0022 -0.1453  0.0557  0.1601
#>  -0.5271 -0.0369  0.1751  0.2188 -0.2555 -0.1406  0.4500 -0.2628  0.1043
#>  -0.5951 -0.2220  0.5008  0.2492 -0.2046 -0.4102  0.0006 -0.2389 -0.0210
#> 
#> Columns 19 to 20 -0.4964  0.1506
#>  -0.4007  0.1736
#>  -0.4418  0.1007
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>