RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4796 -0.5487  0.4954 -0.3612 -0.0935 -0.5106 -0.3434  0.4029 -0.0514
#>  -0.1487  0.6291 -0.1244  0.2441 -0.1720 -0.3947 -0.5272 -0.3724  0.4206
#>   0.7958 -0.6477 -0.0123  0.8295 -0.7517  0.6627  0.5256  0.8817  0.4193
#> 
#> Columns 10 to 18  0.6234 -0.4121  0.2250  0.0655 -0.4424  0.9310 -0.1030 -0.7223  0.4136
#>  -0.4769  0.2834 -0.1202  0.6210  0.3956  0.2822 -0.2124  0.0381 -0.6212
#>   0.9292 -0.2136  0.6364  0.7429 -0.6826  0.1868 -0.8014 -0.7034  0.1545
#> 
#> Columns 19 to 20  0.0635  0.1034
#>   0.0288  0.1887
#>  -0.5336 -0.7305
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1593 -0.1157 -0.6537  0.1815 -0.4596 -0.1730 -0.2191  0.2328  0.6860
#>   0.0867  0.1820  0.4704  0.5352  0.2008 -0.1551 -0.3346  0.0606  0.1576
#>  -0.0070 -0.1750  0.0080 -0.0687 -0.2514 -0.5066 -0.4869 -0.0306  0.7774
#> 
#> Columns 10 to 18 -0.2915  0.2145  0.2637  0.3781 -0.2192  0.1360  0.3852  0.0199  0.3555
#>   0.6030 -0.5420 -0.4435  0.3804 -0.4721  0.7086 -0.2069 -0.4892  0.1102
#>  -0.0818 -0.4626 -0.1559  0.7558 -0.1384  0.0317 -0.4590  0.0312  0.0736
#> 
#> Columns 19 to 20 -0.1409 -0.6022
#>   0.4203  0.0753
#>  -0.3268 -0.5227
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1217  0.1153 -0.0046  0.4996 -0.1071 -0.5740 -0.6542 -0.2994  0.7318
#>  -0.3129  0.0074 -0.0314  0.0299 -0.3204 -0.4359 -0.4537  0.1647  0.2389
#>   0.1581 -0.3483  0.1445  0.3407  0.0505 -0.2449 -0.4857  0.3521  0.6207
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0600 -0.6498  0.5296  0.1517 -0.2445 -0.2269 -0.0712 -0.1877 -0.2628
#>   0.9059 -0.4358  0.5403 -0.0683  0.3122  0.1453  0.3191 -0.3378 -0.2733
#>   0.1173 -0.7335  0.4365 -0.1461  0.1551 -0.4486  0.3348  0.6145 -0.2239
#> 
#> Columns 10 to 18  0.1623  0.0973 -0.1229  0.2930  0.0551  0.1507 -0.1798  0.4459  0.1891
#>   0.1956 -0.1970 -0.3713  0.2460  0.5383  0.8361  0.0774  0.3175 -0.1369
#>  -0.1793  0.0648  0.2000  0.0479  0.3951 -0.1734 -0.0477  0.8252  0.3493
#> 
#> Columns 19 to 20 -0.6087  0.2869
#>  -0.0748  0.5663
#>   0.1498  0.1928
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0684 -0.1343 -0.2892  0.2571 -0.3820 -0.2818 -0.1645  0.1592  0.6301
#>  -0.0038 -0.3569  0.0580 -0.2479 -0.0857 -0.2597 -0.3646  0.3274  0.1504
#>   0.0980 -0.3558 -0.0958  0.2811 -0.2150 -0.3651  0.0069  0.0938  0.5690
#> 
#> Columns 10 to 18  0.2751 -0.2298  0.1222  0.5313  0.0544  0.3907  0.1486 -0.1578  0.1950
#>   0.1347  0.2990  0.0522  0.7162 -0.0950  0.3288  0.1890 -0.5540  0.0262
#>   0.4972 -0.3549  0.2322  0.3408  0.0288  0.5019 -0.0301 -0.4106  0.3034
#> 
#> Columns 19 to 20 -0.0940 -0.4147
#>   0.1706 -0.2696
#>  -0.1534 -0.2161
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>