RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5721  0.4713  0.1135  0.6237  0.2904 -0.7202  0.7243 -0.7227  0.7559
#>   0.6627  0.3514 -0.4104 -0.0865 -0.4169  0.0667 -0.2782 -0.2823  0.5191
#>   0.2818  0.7332 -0.6838  0.6894  0.2608 -0.4118  0.1411  0.0414  0.8898
#> 
#> Columns 10 to 18 -0.1514  0.2730  0.5641  0.2475 -0.0749  0.3234 -0.2438  0.6830  0.8349
#>   0.3562  0.1182  0.1069 -0.7888  0.0948  0.8689  0.5557 -0.3250 -0.3683
#>  -0.9366  0.5060  0.8837 -0.6648  0.3179  0.8246 -0.6828 -0.3239  0.2305
#> 
#> Columns 19 to 20 -0.8171  0.4723
#>   0.3695 -0.6870
#>  -0.4082 -0.6867
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4331 -0.5352  0.0845  0.1972  0.2242  0.5035  0.5382 -0.3625  0.1428
#>   0.4119  0.3415 -0.5895  0.2354  0.6095  0.5572 -0.0561  0.1518  0.3488
#>   0.4668 -0.3577 -0.3168  0.0346  0.7826  0.7190  0.5992 -0.4601  0.6507
#> 
#> Columns 10 to 18  0.5866  0.1659  0.2377  0.0565 -0.1234  0.2372 -0.3798  0.1568 -0.1118
#>  -0.3008 -0.0941  0.2873 -0.5446 -0.0627 -0.1030 -0.2910 -0.3738 -0.3199
#>   0.2880 -0.1583  0.0295 -0.7052 -0.1274 -0.0408  0.3130 -0.0673 -0.4641
#> 
#> Columns 19 to 20 -0.6802 -0.3800
#>   0.1422 -0.0781
#>  -0.2326  0.2129
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5105 -0.1232 -0.1651  0.3494  0.3207  0.0520  0.2243 -0.4969  0.2242
#>   0.0569 -0.2439 -0.1518  0.6591  0.5750  0.3609 -0.2732  0.2786  0.6769
#>   0.2698 -0.2044  0.1488  0.5739  0.5891  0.4557  0.1028  0.3900  0.2667
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7970 -0.3030  0.8601  0.0184 -0.3647  0.4795  0.5578  0.4093 -0.5487
#>  -0.8248  0.2408  0.4303 -0.0102 -0.2808  0.6174 -0.4182  0.2078  0.5904
#>  -0.3941 -0.0469  0.2392 -0.1485  0.6684  0.5618  0.1631 -0.1129  0.0317
#> 
#> Columns 10 to 18  0.7412  0.0728  0.7928  0.2910 -0.6622  0.7873 -0.5606 -0.7751 -0.2658
#>   0.5227  0.6764  0.5836 -0.7452 -0.1207  0.7306 -0.4394 -0.6548 -0.1402
#>   0.3129  0.2146  0.6591 -0.0855 -0.6180  0.2507  0.0643 -0.2454 -0.2560
#> 
#> Columns 19 to 20  0.1682 -0.5186
#>   0.3565 -0.1595
#>   0.3163 -0.6246
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1663 -0.3067  0.4477  0.3547  0.3843  0.7865 -0.1014  0.3080  0.6063
#>   0.5405 -0.3400 -0.2105  0.4478  0.5610  0.4615  0.2732 -0.0707  0.8042
#>   0.2310 -0.0842  0.1295  0.4574  0.3009  0.2205 -0.1385 -0.1055  0.8114
#> 
#> Columns 10 to 18  0.8474 -0.5493 -0.4102 -0.3859 -0.4243  0.0793  0.0413 -0.2213 -0.4103
#>   0.3326 -0.7224  0.3406 -0.6627 -0.0693  0.0013  0.6791 -0.2507  0.0650
#>  -0.2677 -0.1418  0.1259 -0.6807 -0.1813  0.2113  0.2818 -0.2372 -0.0246
#> 
#> Columns 19 to 20 -0.2246 -0.3900
#>  -0.1383  0.2388
#>   0.0393 -0.4036
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>