RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3217  0.7726 -0.7232 -0.4059 -0.3998  0.3958 -0.4056  0.3966  0.1381
#>  -0.8471 -0.7875  0.4390 -0.2457  0.2298 -0.0691 -0.6364  0.5024  0.1712
#>   0.1717  0.6629 -0.1014 -0.3142 -0.2617  0.5065  0.6459 -0.0721  0.3393
#> 
#> Columns 10 to 18 -0.3371  0.6899  0.6817 -0.1672 -0.1993  0.3855  0.0644 -0.4064 -0.5833
#>   0.3619  0.3984 -0.7577 -0.6374  0.2385 -0.9688 -0.6295 -0.9203  0.6952
#>  -0.5585 -0.7155  0.4053 -0.5687  0.0384 -0.7724 -0.2180  0.7487 -0.7575
#> 
#> Columns 19 to 20 -0.3741  0.0923
#>  -0.8893 -0.3431
#>   0.7139  0.1060
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4986 -0.0187  0.2341  0.0274 -0.2484 -0.4027  0.1060  0.3484  0.0683
#>   0.1249 -0.1943  0.0814 -0.6631 -0.3585 -0.6774 -0.1276 -0.6178  0.3924
#>  -0.1473  0.6004 -0.3676  0.1421  0.0091  0.3887  0.7770 -0.5248 -0.2283
#> 
#> Columns 10 to 18 -0.1131  0.5755 -0.2913 -0.1462  0.5334 -0.6679  0.2914 -0.6477  0.2503
#>   0.1334 -0.1259  0.5119 -0.1759  0.8519 -0.3340 -0.0740 -0.4504  0.2224
#>  -0.0654 -0.4763 -0.0733 -0.6133 -0.2928  0.3154 -0.4512  0.1319 -0.6447
#> 
#> Columns 19 to 20 -0.1895  0.0080
#>  -0.2161 -0.1062
#>   0.7267 -0.3108
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3193  0.1015  0.0457 -0.5439 -0.1479  0.1156  0.4356 -0.1661 -0.2309
#>   0.0978 -0.0393  0.1804 -0.2203 -0.5157 -0.2739 -0.1264 -0.0573  0.5315
#>  -0.1747  0.5072 -0.3728 -0.2212 -0.1691  0.6912  0.6950 -0.0451 -0.4000
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0243 -0.8396 -0.4485  0.9191 -0.0913  0.3063 -0.2462  0.4950  0.4999
#>  -0.6065  0.2763 -0.5600 -0.7617  0.4154 -0.0947  0.5392 -0.5766  0.4418
#>  -0.3574  0.4064 -0.0910  0.1948  0.1228 -0.1404  0.5761  0.1212  0.1711
#> 
#> Columns 10 to 18 -0.0937  0.8844 -0.1594 -0.0194  0.7784 -0.0893 -0.7138  0.2958 -0.4825
#>   0.0319  0.4064 -0.4702 -0.6300 -0.5982 -0.7425 -0.2083  0.6607  0.4313
#>  -0.2605  0.0670 -0.2490  0.8230 -0.0170  0.0947 -0.1185  0.4548 -0.3311
#> 
#> Columns 19 to 20 -0.3863 -0.5531
#>   0.3803  0.4824
#>  -0.0295  0.4360
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2434  0.2963 -0.1445 -0.3419 -0.1826 -0.3294  0.3415 -0.5234 -0.0319
#>  -0.2975  0.4443 -0.0540 -0.2008 -0.5075  0.0494  0.0638  0.4352  0.3192
#>   0.0542  0.3449 -0.4164 -0.4879 -0.0631  0.4332  0.3973 -0.1258 -0.2162
#> 
#> Columns 10 to 18 -0.2032 -0.0594  0.2161 -0.2969  0.5486 -0.1735  0.0665 -0.5459 -0.0256
#>   0.6469 -0.0638  0.1880 -0.3833  0.6976 -0.3726  0.0903 -0.3154  0.1266
#>  -0.0558  0.3302  0.3224 -0.5410  0.3741  0.1224 -0.4388 -0.0943 -0.4744
#> 
#> Columns 19 to 20  0.3169 -0.4101
#>  -0.1397 -0.5588
#>   0.4598  0.2638
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>