RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5945 -0.6715  0.0538 -0.1939 -0.5643 -0.4011  0.4678 -0.5292  0.1238
#>   0.3483  0.0718 -0.0381  0.1194 -0.1554  0.1270 -0.1488 -0.8016 -0.2967
#>   0.5936 -0.2669  0.3739 -0.9136  0.2403 -0.1893  0.0994 -0.1825 -0.8756
#> 
#> Columns 10 to 18 -0.3607  0.7024  0.4034  0.3767  0.5010  0.2257  0.5297 -0.3499 -0.2675
#>   0.9235 -0.1206  0.6561 -0.6141  0.4763 -0.2462 -0.5411 -0.3198 -0.1761
#>   0.5586 -0.6724  0.3010  0.4576  0.1093 -0.8159  0.1629  0.0115 -0.0631
#> 
#> Columns 19 to 20  0.5361 -0.3905
#>   0.4185  0.1084
#>   0.2602 -0.3127
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4484 -0.1969 -0.6467 -0.5704  0.5307 -0.0862  0.0725 -0.4592 -0.7591
#>  -0.1548  0.2381 -0.0309 -0.8765  0.5450 -0.0975  0.2754 -0.0878 -0.7230
#>  -0.0431 -0.2228 -0.4043 -0.5440  0.2845  0.3096 -0.1950 -0.6226 -0.6287
#> 
#> Columns 10 to 18  0.4628 -0.4785 -0.3378  0.4783  0.3701  0.0631 -0.4280  0.2299 -0.0523
#>   0.1837 -0.5446  0.3629  0.7676  0.2656 -0.1602 -0.0796  0.5046  0.2754
#>   0.6410 -0.4596 -0.1301  0.4108  0.1244  0.2809  0.0488 -0.6518  0.1422
#> 
#> Columns 19 to 20 -0.1204 -0.6326
#>   0.4397 -0.5263
#>  -0.4793 -0.8281
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0775  0.2294 -0.3654 -0.7119  0.5430 -0.4980 -0.1315 -0.3365 -0.4857
#>   0.3636  0.4515 -0.4007 -0.5180  0.2651 -0.4855  0.0667 -0.4960 -0.4176
#>   0.4148  0.3417 -0.4327 -0.6921  0.6127 -0.5062  0.0478 -0.5500 -0.4897
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6621  0.1229  0.1185 -0.2525  0.0243  0.1674 -0.8605 -0.2567 -0.0350
#>   0.4693 -0.6324  0.1183 -0.3286  0.1198  0.4552 -0.5820 -0.3111 -0.3592
#>  -0.9308  0.5901 -0.2130  0.1859 -0.7486 -0.8015  0.5367  0.6860 -0.4528
#> 
#> Columns 10 to 18 -0.7592 -0.6282  0.2598 -0.5764 -0.1520  0.4831  0.4149  0.0416  0.0316
#>  -0.3168  0.2817  0.4338 -0.2718 -0.0665  0.0709  0.1283 -0.1600  0.1360
#>  -0.3923 -0.7709  0.4656 -0.0953  0.2941 -0.3629  0.6063  0.1812 -0.2642
#> 
#> Columns 19 to 20 -0.0113  0.3228
#>  -0.2709  0.6764
#>  -0.9216  0.6969
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4673 -0.1573 -0.4423 -0.4770  0.2934 -0.3892  0.3011 -0.4991 -0.2438
#>   0.2647 -0.0511 -0.1485 -0.5497  0.4050 -0.1332  0.3714 -0.4516 -0.2732
#>   0.1228  0.0874 -0.8104 -0.5811  0.6396 -0.4494 -0.4017 -0.5589 -0.3246
#> 
#> Columns 10 to 18 -0.6348 -0.6397 -0.3238 -0.0295  0.0914  0.0190  0.6340  0.4138  0.4930
#>  -0.2996 -0.4071 -0.2160 -0.0999 -0.0960 -0.3327  0.5671  0.3450  0.4820
#>  -0.0929 -0.6235 -0.0550 -0.3772  0.2988  0.6973  0.3220  0.1506 -0.3095
#> 
#> Columns 19 to 20  0.2931 -0.5431
#>   0.3093 -0.2337
#>  -0.1322 -0.7340
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>