RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7312 -0.2829  0.4501  0.1720  0.1521  0.4771 -0.8888  0.1214 -0.1314
#>   0.0686  0.3600 -0.0631 -0.4165  0.4422  0.0224 -0.4369  0.6600 -0.1347
#>  -0.4765 -0.0198 -0.8730  0.7132 -0.1346 -0.3753 -0.4040 -0.2488 -0.9010
#> 
#> Columns 10 to 18  0.1471 -0.3388  0.7659  0.1228  0.1181 -0.0335  0.6949  0.7933 -0.5888
#>   0.0836 -0.6449  0.0248 -0.6174 -0.3946 -0.1889  0.5120  0.9277 -0.6067
#>  -0.4516  0.5493 -0.5249 -0.0915 -0.4753 -0.7123  0.0268  0.1238 -0.6104
#> 
#> Columns 19 to 20 -0.0743  0.1631
#>   0.3080  0.0286
#>  -0.4287  0.6130
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5401  0.0432 -0.2807  0.5443  0.0054  0.1937 -0.1836 -0.4092  0.2524
#>   0.0715 -0.0051  0.0847  0.3209 -0.1606 -0.1100 -0.0322 -0.5849 -0.0605
#>   0.1453  0.8091 -0.5889  0.8754 -0.2352 -0.3354 -0.3342 -0.4383 -0.4860
#> 
#> Columns 10 to 18 -0.5907  0.0052  0.6409  0.3524  0.1731  0.1197  0.1045  0.0399  0.2198
#>  -0.4784  0.3888  0.2725  0.0263 -0.2444  0.2381 -0.1320  0.2877  0.1292
#>  -0.2972 -0.1435  0.6399 -0.1656  0.2101 -0.0716  0.3123 -0.5518  0.1229
#> 
#> Columns 19 to 20  0.3219 -0.1458
#>  -0.0518  0.1592
#>   0.0692  0.1391
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1601 -0.1673  0.2620  0.2598 -0.0322 -0.2208 -0.4945 -0.0926  0.0193
#>  -0.1388  0.1845  0.0616  0.0770 -0.1249 -0.1321 -0.2881 -0.4060  0.0076
#>  -0.5469  0.4599 -0.3554  0.1505  0.4106 -0.5534 -0.7566 -0.0446 -0.3451
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2692 -0.2148  0.3318  0.3888  0.5879  0.3448  0.0643 -0.4568 -0.6091
#>   0.1785  0.4263  0.4761 -0.6102  0.3965 -0.5334  0.2695 -0.3550 -0.6361
#>   0.1002  0.1419  0.4171 -0.3194  0.4021  0.3262  0.4018 -0.5291 -0.3634
#> 
#> Columns 10 to 18  0.2452  0.0974 -0.1379  0.4253 -0.1015 -0.2742  0.1577 -0.3863  0.2815
#>  -0.1215 -0.1167 -0.6717  0.4866 -0.0093 -0.1500 -0.0650 -0.3111 -0.7466
#>  -0.0108 -0.1935 -0.6954  0.5586  0.1639 -0.4248 -0.2077  0.6075 -0.6007
#> 
#> Columns 19 to 20  0.4011 -0.7104
#>  -0.4936  0.2284
#>  -0.1342  0.2909
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6413  0.2740 -0.2159  0.2034  0.5530 -0.4348  0.0065 -0.0586 -0.1453
#>  -0.3995 -0.0932 -0.2171  0.1240 -0.6494 -0.2399 -0.3638 -0.5288 -0.2847
#>  -0.1742  0.0028 -0.2224  0.6529 -0.2117 -0.1698 -0.3578  0.0058 -0.3857
#> 
#> Columns 10 to 18 -0.3484  0.2455  0.1479  0.5044  0.0988 -0.1906 -0.0646 -0.0497 -0.1249
#>  -0.4940 -0.2780  0.2773  0.4034  0.0369  0.1804  0.3546 -0.4886  0.3471
#>  -0.4080  0.0756  0.3967  0.5846  0.4662 -0.1127  0.4609  0.0707 -0.0282
#> 
#> Columns 19 to 20  0.0335  0.0239
#>  -0.1516 -0.0805
#>   0.4471 -0.1188
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>