RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3969  0.6373  0.9035 -0.7498 -0.8129 -0.0969  0.2838  0.8632 -0.1160
#>   0.3709 -0.6209  0.2725  0.2265 -0.8994  0.7865  0.2600 -0.1578 -0.4268
#>   0.8585  0.7796 -0.3819 -0.3905 -0.4807  0.1432 -0.0374  0.3702  0.1189
#> 
#> Columns 10 to 18 -0.6959  0.0464 -0.6149 -0.0602 -0.9092  0.0498  0.1028 -0.2974 -0.1193
#>  -0.2888  0.1593 -0.2337 -0.3331 -0.5085 -0.8691  0.8018  0.1687  0.4639
#>   0.1200 -0.7820 -0.2548  0.8756  0.3182  0.6434 -0.1846 -0.1153  0.6769
#> 
#> Columns 19 to 20  0.4776 -0.2826
#>  -0.7268 -0.8754
#>   0.5360 -0.4565
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0243 -0.0931 -0.2333  0.4323  0.1141  0.2573  0.1529 -0.5485 -0.2108
#>   0.6656 -0.4145  0.1807  0.3156  0.0362 -0.3646  0.1731 -0.0833  0.4615
#>   0.3617 -0.4223  0.6752  0.0018 -0.1492  0.0083  0.4036 -0.0295 -0.1336
#> 
#> Columns 10 to 18  0.3763 -0.3380  0.5052  0.4172 -0.4494  0.2406  0.0465 -0.2319  0.1338
#>   0.4142 -0.2301 -0.1435 -0.0611 -0.2341 -0.6416 -0.0136 -0.4708  0.2400
#>  -0.1720 -0.0270  0.1670  0.8063 -0.4286 -0.0735  0.5920  0.0871 -0.2982
#> 
#> Columns 19 to 20  0.2579 -0.2725
#>   0.5002  0.0143
#>   0.5323 -0.3938
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4999  0.0743  0.3121  0.1581  0.2185 -0.0583  0.3241  0.2020  0.0570
#>   0.5223 -0.4311 -0.3331  0.1068  0.0646 -0.0234  0.4457 -0.1977 -0.1248
#>   0.4398  0.4687  0.2984 -0.2242 -0.0323 -0.3140  0.3297 -0.4415  0.0888
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1594  0.2179 -0.1702 -0.0062  0.3147 -0.1142  0.3630  0.4950 -0.2280
#>   0.4713  0.1188 -0.4142 -0.2116 -0.1226 -0.1209 -0.1505 -0.6207  0.3419
#>   0.0589  0.3104 -0.7550  0.4294  0.6999 -0.3448  0.5161  0.2875  0.7507
#> 
#> Columns 10 to 18 -0.1553 -0.2448  0.3660 -0.4072  0.1097  0.4786  0.0083 -0.0830  0.2904
#>  -0.6175 -0.2895 -0.1452  0.4357  0.4336 -0.1813 -0.6880  0.6145  0.1669
#>  -0.1303 -0.6816  0.3807  0.3597  0.2355  0.3218 -0.2285  0.2211  0.3364
#> 
#> Columns 19 to 20  0.4105 -0.1503
#>   0.1789  0.2563
#>   0.1557 -0.8207
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2614  0.1761  0.4774 -0.1275  0.0848 -0.1637  0.4796  0.2140 -0.0438
#>   0.1492  0.2145  0.5141 -0.0434 -0.2697  0.1788  0.0256 -0.0338  0.3892
#>  -0.1386  0.5509  0.3887  0.0704 -0.1313 -0.4281  0.6183  0.2436 -0.3424
#> 
#> Columns 10 to 18 -0.3032 -0.2490  0.1197  0.5928 -0.3043  0.3101  0.1696  0.0436 -0.1433
#>   0.3987 -0.1107  0.5950  0.2125 -0.1107  0.3234  0.1008  0.0310  0.5011
#>  -0.0759 -0.1984  0.2387  0.3719 -0.4813  0.1489  0.3815  0.2573 -0.1547
#> 
#> Columns 19 to 20  0.0732 -0.0868
#>  -0.0588 -0.1159
#>  -0.3479 -0.1879
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>