RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3731 -0.4215 -0.1186  0.5088  0.4968 -0.4478 -0.2065 -0.2007 -0.8394
#>   0.2029 -0.3447 -0.3546 -0.3192 -0.1936  0.2580  0.8664 -0.6663 -0.0736
#>   0.5289  0.3731 -0.8132  0.8526 -0.7381 -0.0273  0.8718  0.4398 -0.8177
#> 
#> Columns 10 to 18 -0.4730 -0.4034 -0.0628 -0.8807  0.4285  0.6539  0.5218  0.1716 -0.8909
#>  -0.2104 -0.6826 -0.1839  0.7704 -0.2697 -0.2868 -0.8191  0.0481 -0.6598
#>   0.5005  0.6607  0.1417 -0.5455 -0.1366 -0.0132  0.2595  0.2872  0.5174
#> 
#> Columns 19 to 20 -0.0905  0.3944
#>   0.3304  0.3979
#>  -0.0000  0.2129
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6773 -0.3956 -0.4145 -0.1933  0.0886  0.3639  0.1837 -0.2429  0.2454
#>   0.6144 -0.2574 -0.5413 -0.0675 -0.1520  0.0806  0.7550  0.3482 -0.0434
#>  -0.4437  0.1657 -0.2760 -0.0675  0.5804 -0.2784 -0.3468  0.1612 -0.1869
#> 
#> Columns 10 to 18 -0.6167  0.3905 -0.1079 -0.2404  0.5189 -0.0830  0.1820  0.5414 -0.2741
#>  -0.3006  0.6400 -0.2247 -0.1049  0.3751 -0.4103 -0.2523 -0.0411  0.1214
#>   0.5958  0.4102 -0.3879  0.0854  0.4232  0.4554 -0.0098  0.4365 -0.0609
#> 
#> Columns 19 to 20 -0.2300 -0.2095
#>   0.0693 -0.0947
#>  -0.2102 -0.1222
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2312 -0.3129 -0.2809  0.4205  0.3218 -0.1265 -0.4173  0.1051 -0.3816
#>  -0.0011 -0.3063 -0.0633  0.2985  0.2859 -0.1999  0.2205  0.3457 -0.2137
#>   0.1774 -0.0647  0.1844  0.1249  0.4743  0.3292 -0.4560  0.0902 -0.3621
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1603  0.7733  0.6898 -0.2787 -0.3852  0.7066  0.1787  0.5148  0.4787
#>  -0.6847 -0.3804 -0.3140 -0.6335 -0.5403 -0.5335  0.1014 -0.5813 -0.8207
#>  -0.7191  0.2656  0.3834 -0.3560  0.3089  0.2484  0.1726  0.5357  0.4385
#> 
#> Columns 10 to 18  0.9179  0.3214 -0.3131  0.0220  0.5845 -0.3227  0.3867  0.4323  0.4544
#>  -0.3389 -0.6698  0.2854 -0.0989 -0.1857  0.4177 -0.7615 -0.5155 -0.4481
#>   0.2685  0.2107 -0.3280 -0.6099  0.3580  0.2776 -0.0206  0.5325 -0.0952
#> 
#> Columns 19 to 20  0.2776 -0.3829
#>  -0.3754  0.0438
#>  -0.0047 -0.7922
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3907 -0.5850 -0.1259 -0.3214  0.5675 -0.0413 -0.3186 -0.0188 -0.4668
#>   0.2067 -0.2229 -0.3244 -0.1870 -0.3510  0.5499  0.6536  0.0631  0.0277
#>   0.4476 -0.0277 -0.2864  0.1704  0.3457 -0.1552 -0.0276 -0.0268 -0.2976
#> 
#> Columns 10 to 18 -0.1204  0.3840  0.1398  0.1410 -0.0889  0.0460 -0.4924  0.3136  0.0754
#>  -0.0501  0.5447  0.3885 -0.2831  0.5514 -0.3752  0.1502 -0.0100  0.1207
#>   0.1341  0.3822  0.2661 -0.2432  0.1015 -0.3901 -0.2814  0.3576  0.3567
#> 
#> Columns 19 to 20 -0.1612  0.2472
#>   0.2853 -0.1802
#>  -0.6316  0.2426
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>