RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6753 -0.0223 -0.1378 -0.6258 -0.8344  0.0972  0.2900 -0.3107  0.8099
#>  -0.6048  0.2636  0.6241 -0.3099 -0.4162 -0.7003 -0.2740 -0.2583  0.1190
#>  -0.1139 -0.7827 -0.5811 -0.6019 -0.9012  0.4225  0.3665  0.6364 -0.8355
#> 
#> Columns 10 to 18  0.4189  0.0272  0.5016 -0.0594 -0.4309 -0.3125  0.4962  0.3706  0.6793
#>   0.8126 -0.6777 -0.5759 -0.8952 -0.4072 -0.6900  0.8197  0.0133 -0.3642
#>  -0.3692  0.5305  0.9431  0.4884  0.3022  0.7184  0.5020 -0.0038  0.0632
#> 
#> Columns 19 to 20 -0.7175  0.5788
#>  -0.0274  0.1026
#>   0.3525  0.4845
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5359  0.3045 -0.3930  0.3341 -0.4518 -0.5502  0.4361 -0.1380 -0.2214
#>  -0.6488 -0.3115 -0.0079  0.0474 -0.4726  0.0502 -0.1834 -0.3258 -0.2483
#>  -0.0048 -0.1179 -0.2565  0.0885 -0.1258  0.1130  0.6423 -0.5160 -0.0179
#> 
#> Columns 10 to 18 -0.0686 -0.0282  0.2475 -0.2302 -0.1745  0.5983 -0.1884  0.2074 -0.5275
#>   0.5509 -0.5878 -0.6884 -0.6010  0.5311  0.0873  0.1115  0.5323 -0.1484
#>   0.5657  0.0392  0.3191 -0.1500  0.4709  0.0423 -0.0245  0.1073 -0.7448
#> 
#> Columns 19 to 20 -0.5923 -0.0395
#>   0.2757 -0.0410
#>  -0.1100  0.2503
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1017  0.1048 -0.4433  0.2029 -0.2733 -0.3282  0.7175 -0.2627 -0.1816
#>  -0.4516 -0.0688 -0.2391 -0.0635 -0.0803 -0.4651 -0.2040  0.0705  0.0734
#>  -0.0364 -0.3002 -0.5551  0.6071 -0.0702  0.4129  0.3135 -0.5753  0.1957
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4697  0.2312  0.1744  0.3484  0.0818 -0.3858 -0.3886  0.1966 -0.1270
#>  -0.1755  0.6199 -0.2129 -0.1137  0.6941 -0.6154  0.3166 -0.5659 -0.1544
#>  -0.1399  0.4359  0.4238  0.7342  0.5847 -0.4667 -0.3337  0.1873 -0.1435
#> 
#> Columns 10 to 18 -0.4770 -0.0479  0.0954 -0.1913 -0.2153 -0.0151 -0.4516  0.8673 -0.1207
#>   0.1697  0.1616  0.6551 -0.3184 -0.4139  0.0837 -0.1796  0.5682 -0.6395
#>   0.0779  0.2542 -0.1159 -0.3580 -0.7914 -0.0175 -0.3566  0.4726 -0.6921
#> 
#> Columns 19 to 20 -0.1138 -0.3113
#>  -0.2176 -0.0597
#>  -0.0481 -0.0280
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1331 -0.1917 -0.6150  0.2871 -0.1479 -0.3751  0.3578  0.0366 -0.3596
#>  -0.3023 -0.0770 -0.1107  0.1485 -0.6217 -0.4412  0.0382 -0.2796 -0.4547
#>   0.1575  0.0433 -0.2466  0.4126 -0.4194 -0.5439  0.0526  0.2640 -0.4837
#> 
#> Columns 10 to 18  0.1835 -0.3147 -0.1842 -0.4143  0.5818  0.3683 -0.2654  0.1849 -0.6182
#>   0.3140 -0.5253 -0.3275 -0.1855  0.2823  0.6280 -0.6342  0.2057 -0.1210
#>   0.2707 -0.2080 -0.2788 -0.2454  0.4053  0.7717 -0.3578  0.0021 -0.4031
#> 
#> Columns 19 to 20 -0.3161 -0.0032
#>  -0.4000  0.2511
#>  -0.4982 -0.3002
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>