RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6953  0.3299  0.8254  0.7776 -0.7723 -0.6343 -0.5006  0.3732  0.5482
#>  -0.2515 -0.5269  0.7647 -0.1986  0.2409  0.0083 -0.1450 -0.1110 -0.8403
#>  -0.5919 -0.4613  0.7409  0.6342 -0.4206  0.1100  0.3145 -0.2316 -0.7891
#> 
#> Columns 10 to 18  0.8390 -0.2043  0.3485 -0.4386 -0.2455  0.0176 -0.8145 -0.0223 -0.1629
#>   0.4841 -0.1494  0.7178 -0.7011  0.0139 -0.1770 -0.3900 -0.3242  0.8697
#>  -0.4406  0.2628  0.9319 -0.3530  0.1832  0.3858 -0.0138 -0.7585  0.2561
#> 
#> Columns 19 to 20 -0.0568 -0.5217
#>  -0.0053 -0.5130
#>   0.5775 -0.2805
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5032  0.5166  0.5709 -0.1819 -0.2537 -0.5295  0.3699 -0.5278 -0.1377
#>  -0.1339  0.2295  0.3811 -0.7189 -0.4139 -0.1775  0.6446 -0.5726 -0.6450
#>  -0.1980 -0.2949  0.2679 -0.5977 -0.5240 -0.1397  0.2008 -0.4323 -0.2173
#> 
#> Columns 10 to 18  0.6742 -0.0424 -0.0580 -0.3791 -0.5657  0.6616  0.3517 -0.1652  0.3501
#>  -0.3959  0.5565 -0.3274  0.3290 -0.6587  0.0615  0.2848  0.2318  0.5714
#>   0.3360  0.0453  0.2934  0.0845 -0.6800  0.2725  0.0848  0.5457  0.4553
#> 
#> Columns 19 to 20  0.2563 -0.4411
#>   0.4150 -0.6107
#>   0.0146  0.0176
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5264 -0.0423  0.3997  0.0125 -0.4238 -0.3555 -0.0379 -0.5599 -0.4198
#>  -0.6499  0.3020 -0.1254 -0.0919 -0.4649 -0.4268  0.4047 -0.3052 -0.2718
#>  -0.6025  0.5539  0.3531 -0.2391 -0.3467 -0.1862  0.2637 -0.5193 -0.3671
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1149 -0.3349 -0.4811 -0.0650 -0.1778  0.0740 -0.6479 -0.3932  0.1840
#>   0.1015 -0.1417 -0.5930 -0.4582 -0.4536  0.3581 -0.3927 -0.4699  0.4141
#>   0.7107  0.2409 -0.1641 -0.8074  0.1375 -0.0108 -0.3343 -0.5204 -0.7548
#> 
#> Columns 10 to 18  0.1197  0.7316 -0.2546  0.4547  0.3164 -0.3493  0.1569 -0.3562 -0.0632
#>   0.3109  0.0364 -0.5097  0.0451 -0.0376 -0.1677 -0.0068 -0.2649  0.6278
#>   0.5163  0.3687  0.2442  0.5238 -0.9034  0.3038 -0.6639 -0.7443  0.0948
#> 
#> Columns 19 to 20 -0.4522 -0.1850
#>  -0.3799 -0.0367
#>  -0.5241 -0.3478
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1817  0.2233  0.5308 -0.1925  0.0004 -0.6055  0.5036 -0.7412 -0.2637
#>   0.1194  0.0836  0.5850 -0.2354 -0.2818 -0.2893 -0.1220 -0.3986 -0.2765
#>  -0.0559 -0.2820  0.6155  0.2604  0.1995 -0.0407 -0.4691 -0.6598 -0.5415
#> 
#> Columns 10 to 18  0.3392  0.1078 -0.4305 -0.2574 -0.1149  0.3769 -0.0999 -0.4813  0.2326
#>   0.5370 -0.1105 -0.2118 -0.2463 -0.5037  0.3290 -0.2404 -0.1319  0.2236
#>   0.2753 -0.0989 -0.1507  0.3041 -0.4136  0.2261  0.0784 -0.3287  0.1979
#> 
#> Columns 19 to 20 -0.3131 -0.1054
#>   0.0940 -0.2961
#>  -0.3373 -0.0886
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>