RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2405  0.7805  0.5148  0.8010  0.7127  0.0024  0.7503 -0.2191  0.5351
#>   0.2082  0.8286  0.5527 -0.8658 -0.2195 -0.6880 -0.5866  0.1670  0.0078
#>  -0.7200  0.3653  0.7337  0.5046  0.4647  0.6506 -0.7008  0.8380 -0.5890
#> 
#> Columns 10 to 18 -0.2827 -0.1705 -0.6151 -0.3804  0.5178 -0.1985 -0.5586  0.3646  0.1527
#>   0.2836  0.4526  0.1230 -0.5377  0.2840  0.0933 -0.0138 -0.3615 -0.4065
#>  -0.5500 -0.4649 -0.5545 -0.5714  0.0536 -0.6918 -0.7655  0.5697 -0.8237
#> 
#> Columns 19 to 20  0.0072 -0.2335
#>   0.2456  0.1210
#>  -0.6366 -0.7704
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4350 -0.3676 -0.4872 -0.2744  0.3902 -0.6973  0.1287  0.5159  0.5580
#>   0.7297 -0.3105  0.5010 -0.5114  0.1413 -0.2799 -0.2852  0.2342  0.3120
#>   0.3941 -0.6649  0.4222 -0.2066  0.0461 -0.1376  0.2618  0.3573  0.4934
#> 
#> Columns 10 to 18 -0.0434  0.1363 -0.3953 -0.7167 -0.7714 -0.0282  0.6077 -0.0408  0.6183
#>   0.1155  0.4430 -0.4747 -0.3462 -0.3584 -0.6970  0.1574  0.1500  0.1177
#>  -0.2486  0.3598 -0.3047 -0.4411  0.2592 -0.0513  0.3772  0.2496 -0.0752
#> 
#> Columns 19 to 20 -0.0646  0.1733
#>  -0.6251 -0.0531
#>   0.4573  0.2571
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2419 -0.1526 -0.0446  0.1436  0.3276  0.0527  0.2759  0.3856  0.2033
#>  -0.2707  0.0230 -0.0137  0.1143  0.3503 -0.2172 -0.3219  0.7224  0.3604
#>   0.1574 -0.1306  0.1662  0.2134 -0.1631  0.0338 -0.0937 -0.0515  0.1711
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1111 -0.1784 -0.0559 -0.3221  0.4788 -0.0147 -0.4595  0.2291 -0.1681
#>   0.4772 -0.6211 -0.2556 -0.0499  0.1907 -0.4846  0.1359  0.1264  0.0682
#>   0.2871 -0.1403  0.1482 -0.1951 -0.0890 -0.1342  0.0913 -0.6032 -0.4802
#> 
#> Columns 10 to 18 -0.6712  0.3069 -0.5958 -0.6576 -0.2182 -0.0698 -0.7647  0.4909  0.5074
#>  -0.3355 -0.1754  0.0082 -0.6239  0.0012  0.4225 -0.3516 -0.0451  0.6318
#>  -0.1819  0.4232 -0.1999  0.1106  0.4212 -0.7707 -0.0735  0.0551  0.1049
#> 
#> Columns 19 to 20 -0.2158  0.3378
#>  -0.2314 -0.1867
#>  -0.5574 -0.0536
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1335 -0.4560  0.1367  0.3920  0.2954 -0.1674  0.0565  0.3623  0.2652
#>  -0.0269 -0.1744  0.3119  0.1741  0.0953 -0.3690  0.0352  0.2406  0.4310
#>  -0.1203  0.2096  0.1889 -0.1448  0.3289  0.3181  0.3875  0.4671  0.1296
#> 
#> Columns 10 to 18  0.1724 -0.0124 -0.2556 -0.6211 -0.3562 -0.4110 -0.0486 -0.1923  0.6826
#>   0.6394 -0.3940 -0.3275 -0.4107 -0.4851  0.0898  0.5203 -0.1824  0.7149
#>  -0.1943  0.0951 -0.1838 -0.6485  0.4805  0.1582 -0.2052  0.3783  0.3432
#> 
#> Columns 19 to 20 -0.0206 -0.2857
#>  -0.5388 -0.0664
#>  -0.4827 -0.1641
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>