RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2140 -0.0732  0.0909 -0.5396 -0.3913 -0.7196  0.3276  0.1971 -0.3059
#>  -0.7392 -0.7284 -0.0164 -0.0008  0.0444  0.1929 -0.2668 -0.6650  0.4093
#>   0.4690  0.2021 -0.6935  0.8268 -0.7074 -0.7255 -0.0608 -0.4473 -0.9320
#> 
#> Columns 10 to 18  0.2130 -0.5522  0.1842 -0.6805  0.7897 -0.4856 -0.7089 -0.8903 -0.1401
#>  -0.3704 -0.3889  0.4071  0.1017  0.7561  0.7344 -0.1692 -0.1276  0.7493
#>  -0.7345 -0.6600  0.0399 -0.4543  0.5261 -0.6342 -0.8136  0.1779 -0.0261
#> 
#> Columns 19 to 20 -0.2254 -0.2292
#>   0.6728  0.7885
#>  -0.1776  0.7943
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6511 -0.1757 -0.3721  0.0366 -0.4220 -0.0232 -0.0689 -0.1095  0.1202
#>   0.2268  0.6770  0.6589 -0.4328 -0.4669 -0.2275 -0.6065  0.0647 -0.1524
#>   0.0528 -0.1323  0.4224 -0.5960  0.0580  0.0154 -0.4070  0.5038 -0.0715
#> 
#> Columns 10 to 18 -0.5160 -0.1133  0.2945  0.0613  0.6022 -0.0084 -0.0912  0.2986 -0.1728
#>  -0.2136 -0.1586  0.7141  0.7684 -0.6416  0.5571  0.1310 -0.3100 -0.4127
#>   0.4782 -0.1074  0.7068  0.1131  0.4447  0.3847  0.2146 -0.3459  0.1791
#> 
#> Columns 19 to 20  0.5167  0.5923
#>   0.7055  0.5576
#>   0.2411  0.0938
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2542  0.0149  0.3692 -0.3662  0.0621 -0.0712 -0.5779 -0.0510 -0.1906
#>  -0.6703 -0.1672  0.1022  0.1971 -0.1974 -0.3949  0.0404 -0.0356  0.0778
#>  -0.1798 -0.4033 -0.3861 -0.2748 -0.4532  0.0586  0.0608 -0.4185  0.0772
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1078 -0.0654  0.5281 -0.4385  0.4469 -0.0406 -0.8699 -0.1812  0.2643
#>   0.2194  0.1525  0.3848  0.5630 -0.0197 -0.2557 -0.8937  0.2326  0.5003
#>  -0.2543 -0.8450  0.5743 -0.4754  0.5359 -0.5323 -0.1255  0.0079 -0.1190
#> 
#> Columns 10 to 18 -0.0677  0.6758  0.0926  0.0494  0.4841 -0.1964  0.0378 -0.1544  0.2016
#>  -0.0562  0.5953 -0.3337 -0.0707  0.2835 -0.5903  0.1902 -0.8175  0.2203
#>   0.1767 -0.2576  0.2110 -0.3338 -0.0253  0.0007 -0.1515  0.7898 -0.1177
#> 
#> Columns 19 to 20  0.0331  0.1367
#>   0.2761 -0.4097
#>   0.6186  0.3203
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0926 -0.1466 -0.0332 -0.2916 -0.3831 -0.1351 -0.5116 -0.0515  0.2999
#>  -0.2036  0.0526  0.4312 -0.3034 -0.2185 -0.1289 -0.5370  0.0636  0.1866
#>   0.1633 -0.4807 -0.2161 -0.2715 -0.3027 -0.1674  0.0417 -0.3076 -0.1185
#> 
#> Columns 10 to 18 -0.2804 -0.1297  0.6154  0.2246  0.2663  0.5338 -0.0269 -0.3633  0.3165
#>   0.0568 -0.2136  0.1796  0.1346  0.1692  0.6507  0.0427 -0.5312  0.3682
#>  -0.4170 -0.4749  0.0687  0.4123  0.4822  0.1388 -0.3769 -0.1712 -0.1912
#> 
#> Columns 19 to 20  0.5571 -0.0376
#>   0.6614 -0.0366
#>   0.4841  0.2241
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>