RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4872  0.3867  0.8050 -0.1293  0.7356 -0.2922  0.8514 -0.6779  0.2141
#>  -0.1533 -0.2564 -0.6390  0.8252 -0.2751  0.4553  0.7830  0.7917 -0.2112
#>   0.6711  0.5865  0.1442 -0.0505  0.3768 -0.4776 -0.5173 -0.3241  0.7411
#> 
#> Columns 10 to 18 -0.5803  0.1299  0.7237  0.6664 -0.3502  0.6118 -0.6794 -0.5584 -0.4093
#>  -0.0537 -0.6335  0.3560  0.8156  0.9311 -0.6525  0.4500 -0.7137  0.0206
#>  -0.4473  0.9484 -0.0359 -0.7281 -0.2927 -0.3800 -0.4272 -0.1144  0.7642
#> 
#> Columns 19 to 20  0.8425 -0.5136
#>   0.9283 -0.7704
#>  -0.5293 -0.8016
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6122 -0.0395 -0.7321  0.0932 -0.1233 -0.2461  0.0103 -0.3997  0.7366
#>   0.7013  0.4354  0.0066  0.1020  0.3112  0.0440  0.1738  0.0938  0.7186
#>   0.3146 -0.3821 -0.0520 -0.2425  0.1162 -0.6999 -0.2403 -0.2394  0.3191
#> 
#> Columns 10 to 18  0.5993 -0.0518 -0.3303 -0.0918 -0.1684 -0.6882 -0.4450 -0.6697 -0.4568
#>   0.6213  0.0716  0.1689  0.3141  0.4785 -0.8226  0.2414 -0.0642 -0.1089
#>   0.2031 -0.0926 -0.3819  0.3222 -0.0581  0.4993  0.0396 -0.1995 -0.0969
#> 
#> Columns 19 to 20  0.1913 -0.2345
#>  -0.1590  0.1757
#>   0.7136 -0.0594
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4309  0.1035  0.3462 -0.1093  0.5137 -0.2149  0.2681 -0.5025  0.4971
#>   0.1970  0.1461  0.1587  0.2095  0.2360  0.1657  0.4628 -0.0303  0.4888
#>   0.2323 -0.5608  0.1548 -0.0166  0.0959 -0.3869 -0.0384 -0.4756  0.5634
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2032 -0.4838  0.7692  0.5525 -0.3226  0.1806  0.7416  0.6068 -0.6962
#>   0.2139 -0.3463  0.8224 -0.0856 -0.0120 -0.0369 -0.0833 -0.2945 -0.6512
#>  -0.1220  0.5818 -0.4588 -0.5674  0.0895 -0.3431 -0.4605  0.0734  0.2047
#> 
#> Columns 10 to 18 -0.0568  0.5835 -0.3109  0.0493 -0.4363 -0.3587  0.5017  0.0956  0.3601
#>  -0.6003 -0.0352  0.4324  0.7016  0.7610 -0.6590 -0.3366  0.2130  0.4417
#>  -0.0498  0.4920  0.8434  0.5195 -0.8026 -0.6230  0.1348 -0.3001  0.5079
#> 
#> Columns 19 to 20 -0.0115  0.3823
#>  -0.5205 -0.4086
#>   0.3978  0.5136
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3055 -0.4057  0.0847  0.0369  0.2249 -0.5396 -0.2500 -0.5071  0.1006
#>   0.6212 -0.3883 -0.0542 -0.0600  0.1440 -0.4967  0.4168 -0.1699  0.4143
#>   0.0410  0.5379  0.0414  0.1403  0.2565  0.4050  0.4599  0.0412  0.6097
#> 
#> Columns 10 to 18  0.4462  0.1035 -0.1475  0.2666  0.1035  0.1829  0.0458  0.1637  0.4976
#>   0.2419  0.3292  0.2806  0.5939  0.0875 -0.5327 -0.4095  0.0535 -0.3852
#>   0.5986 -0.1101  0.1192  0.5857  0.6043 -0.5876 -0.1367  0.0457  0.5064
#> 
#> Columns 19 to 20  0.4375 -0.0262
#>   0.3005  0.4329
#>   0.1211 -0.5515
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>