RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1725  0.3809  0.6357 -0.1042 -0.7300  0.7270  0.4674 -0.5773 -0.7019
#>   0.0604 -0.3137  0.8047  0.8055 -0.9014  0.8850 -0.7925 -0.9509 -0.6150
#>  -0.0981  0.5273 -0.5411 -0.1949  0.6545 -0.8718 -0.8761 -0.9065 -0.9190
#> 
#> Columns 10 to 18  0.6005  0.5855  0.4979  0.4413 -0.1636 -0.4688  0.6176  0.1962 -0.7837
#>  -0.8701 -0.5246 -0.7255  0.3047 -0.6921 -0.3717  0.3797  0.9008 -0.7842
#>   0.5516 -0.1606 -0.9802 -0.3504 -0.3229  0.2088  0.3377  0.3158  0.6194
#> 
#> Columns 19 to 20 -0.6858  0.3288
#>  -0.9279 -0.3586
#>   0.7571  0.8600
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6364  0.4635 -0.0425 -0.0339  0.2411 -0.4828  0.0351 -0.5710  0.2231
#>  -0.8508  0.3338  0.6385  0.3249 -0.7267 -0.1341 -0.1030 -0.5192  0.3698
#>  -0.2814 -0.4696 -0.1064 -0.5187  0.3940  0.4559  0.0321 -0.4013 -0.3188
#> 
#> Columns 10 to 18 -0.1878 -0.4921  0.1935  0.4565 -0.3672 -0.6172  0.0623  0.2101  0.2070
#>  -0.3259  0.0155  0.7393  0.6960 -0.7033  0.2380 -0.3497 -0.4896 -0.2364
#>   0.2248  0.4076  0.1624  0.0458  0.1576  0.0917 -0.2782  0.2272  0.2258
#> 
#> Columns 19 to 20  0.4651  0.7100
#>   0.1614  0.4346
#>  -0.7734 -0.0136
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0128 -0.1068 -0.1283 -0.0871  0.1151  0.2641 -0.2193 -0.3767  0.1507
#>  -0.4789  0.2806  0.0781  0.3061 -0.3708 -0.1455  0.0491 -0.5816 -0.3543
#>  -0.4754 -0.1593  0.4286  0.5237 -0.0842  0.0134 -0.0980 -0.5839  0.3884
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0699  0.3516 -0.0671  0.6435 -0.5445  0.1930  0.7039 -0.4525 -0.1935
#>  -0.3531 -0.0993  0.0705 -0.1527 -0.0623 -0.3047  0.2345 -0.1417 -0.3424
#>  -0.1508  0.3538 -0.0194 -0.4614 -0.1028 -0.0008  0.0775  0.0535  0.0431
#> 
#> Columns 10 to 18  0.2227  0.2877 -0.3671 -0.3532  0.0040  0.6811  0.1756 -0.2017  0.7189
#>   0.0256 -0.8026 -0.2771 -0.3462 -0.5992  0.1896  0.4111 -0.0979  0.3888
#>  -0.4809  0.0935 -0.0525 -0.1470 -0.3265  0.0980  0.0330 -0.0228  0.2191
#> 
#> Columns 19 to 20  0.1033 -0.7150
#>   0.0191  0.3406
#>  -0.7515 -0.2132
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1332 -0.2863  0.1300  0.0182 -0.2063 -0.0384 -0.0222 -0.7783 -0.0012
#>  -0.3013 -0.1600  0.0302  0.1057 -0.0860 -0.1950 -0.1295 -0.0761  0.4255
#>  -0.3096  0.0453 -0.1274 -0.0024  0.1821  0.0525  0.0352 -0.3375  0.1812
#> 
#> Columns 10 to 18 -0.2142 -0.0952  0.0866  0.2863 -0.4303 -0.2095  0.2610  0.4505 -0.1837
#>  -0.1555 -0.0498  0.1422  0.4280 -0.3160 -0.0289 -0.3058  0.3766  0.1004
#>  -0.0912 -0.1606  0.1964  0.2547 -0.0153 -0.2305  0.0889  0.2137  0.2643
#> 
#> Columns 19 to 20 -0.2891  0.5072
#>   0.1205  0.1276
#>   0.0734  0.5038
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>