RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2082 -0.2485 -0.1703 -0.2715 -0.2070  0.1638 -0.3402  0.2408 -0.6327
#>   0.0580  0.7088 -0.2619  0.0086  0.7979 -0.8718 -0.6461  0.0381  0.1809
#>   0.4323  0.1985  0.5601 -0.3986  0.2131  0.8629  0.6182  0.4130 -0.0866
#> 
#> Columns 10 to 18  0.1394 -0.7740  0.1996  0.7146  0.2178  0.5094  0.0145  0.8524  0.5821
#>   0.1428 -0.3748  0.5133  0.7486 -0.6795  0.5768  0.0187  0.7451 -0.3042
#>  -0.0365 -0.6590 -0.9336 -0.4556  0.8118  0.6848 -0.8184  0.1223  0.5409
#> 
#> Columns 19 to 20 -0.3808  0.0729
#>   0.5886  0.8626
#>  -0.9372 -0.2570
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0432 -0.6060 -0.0235  0.0955  0.1745 -0.3579  0.0211 -0.1839  0.0748
#>   0.0267 -0.5145  0.0979  0.6051  0.4155 -0.6470 -0.1425  0.2586 -0.1053
#>   0.8299  0.1198  0.3266 -0.1668  0.6882  0.0796  0.3835 -0.0192  0.3029
#> 
#> Columns 10 to 18 -0.5531  0.1050  0.2862  0.0918  0.6529 -0.0433 -0.2302  0.1572 -0.4917
#>  -0.6746 -0.0148  0.2163  0.0491  0.0821 -0.2563  0.2657  0.5114 -0.3924
#>   0.2834 -0.5352 -0.0412  0.3245  0.6400 -0.6250  0.1531  0.8323  0.2995
#> 
#> Columns 19 to 20 -0.5694 -0.1869
#>  -0.0546  0.6970
#>  -0.3686 -0.4572
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3239 -0.2937  0.4494  0.2790  0.4841  0.1568  0.3479 -0.3744  0.0467
#>   0.6079 -0.3439  0.1740  0.0315  0.2772 -0.4515  0.3868 -0.0007  0.5086
#>   0.5040 -0.0877  0.5088  0.6034  0.2008 -0.1482  0.5778  0.1553 -0.1498
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4132 -0.5093 -0.0786 -0.3160 -0.6995 -0.5090  0.0699  0.1143  0.1265
#>   0.1385 -0.5832 -0.5766  0.0817  0.1209 -0.5319 -0.0654  0.0559 -0.1405
#>   0.0157 -0.1437 -0.4012  0.0403 -0.3027  0.0205 -0.3686  0.0760 -0.4026
#> 
#> Columns 10 to 18 -0.3061 -0.4855 -0.1254  0.4021 -0.0531 -0.1202 -0.2266 -0.0650 -0.1540
#>  -0.2928 -0.3810  0.2344  0.6371 -0.5441 -0.3766  0.6253  0.1065  0.3333
#>   0.0645 -0.4162  0.1821  0.4701  0.5660  0.1788  0.5819  0.2202  0.6531
#> 
#> Columns 19 to 20 -0.5872 -0.5409
#>   0.4586 -0.7894
#>  -0.4913 -0.6930
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5633 -0.2865  0.1435  0.4523  0.0364 -0.4255  0.5847  0.0988  0.1681
#>   0.5706 -0.3379  0.4940 -0.1923  0.1990 -0.3546  0.4117 -0.1191 -0.0687
#>   0.5543 -0.2961  0.0789  0.1117 -0.2511 -0.5222  0.1808  0.2291  0.1696
#> 
#> Columns 10 to 18 -0.5054 -0.5469 -0.1378 -0.2286  0.0085  0.0537 -0.4497  0.5116 -0.1104
#>  -0.1490 -0.8835  0.0013  0.0025 -0.2068  0.0805 -0.4249  0.4544 -0.4469
#>  -0.5107 -0.5766 -0.3602 -0.1739 -0.0928  0.4889 -0.4678  0.3638 -0.2247
#> 
#> Columns 19 to 20 -0.2148 -0.1100
#>  -0.0873  0.0014
#>  -0.2411 -0.1958
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>