RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0345 -0.3599  0.1554  0.6968 -0.3848 -0.5169 -0.5517 -0.6118  0.8221
#>  -0.8865  0.9575  0.8311 -0.4249  0.6232 -0.5756  0.4471  0.2107 -0.3483
#>  -0.5998  0.2221  0.7780 -0.8352  0.2519 -0.4196 -0.6073 -0.6799  0.4748
#> 
#> Columns 10 to 18 -0.5021 -0.4260  0.4943  0.4852  0.7153 -0.0269  0.4925  0.3962  0.6335
#>  -0.3438 -0.5593 -0.3933 -0.0199 -0.8975 -0.7473 -0.1454  0.8535  0.5161
#>  -0.1233 -0.2016  0.6976  0.7513 -0.3917  0.3163 -0.3852 -0.7319  0.5841
#> 
#> Columns 19 to 20  0.1935 -0.3817
#>  -0.5799  0.1966
#>  -0.2803 -0.3781
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3806  0.5114 -0.0674 -0.4547 -0.1751 -0.5685 -0.4823  0.4256 -0.4934
#>   0.4423  0.3258 -0.1983 -0.1124 -0.3319 -0.5385 -0.7927 -0.4351 -0.3427
#>   0.3149 -0.0160  0.0180 -0.8524  0.3009 -0.2725 -0.5408  0.0978 -0.4956
#> 
#> Columns 10 to 18  0.1335 -0.3801  0.3000 -0.6912  0.4374  0.3734 -0.2267 -0.3241 -0.5315
#>   0.6216 -0.3196  0.7585  0.2235  0.3264  0.2911  0.0897 -0.1049  0.0046
#>   0.5329 -0.1190 -0.1889  0.0145  0.2050  0.1586  0.0682  0.6565 -0.3830
#> 
#> Columns 19 to 20  0.2509  0.1953
#>   0.0179  0.0205
#>  -0.3188  0.2296
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5638  0.2051  0.3577 -0.0218 -0.4660 -0.0345 -0.3073 -0.6438 -0.2204
#>  -0.6554  0.6053  0.1160 -0.1302 -0.2860 -0.5948 -0.0215 -0.1568  0.1490
#>  -0.3312 -0.2095  0.2461  0.5191  0.0784 -0.4231 -0.0561 -0.5049 -0.1000
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2015  0.0569  0.0657  0.7437 -0.6931  0.4831  0.0054  0.2279  0.0692
#>  -0.1666  0.1891 -0.1012  0.4710 -0.0921 -0.3282  0.5499 -0.2041  0.3905
#>   0.3219 -0.3817  0.1790  0.2822 -0.2366  0.1689 -0.2572 -0.2161 -0.0630
#> 
#> Columns 10 to 18 -0.5568 -0.7818 -0.0810  0.7171  0.7307  0.2887 -0.5075 -0.1614  0.3973
#>   0.4066 -0.5213 -0.4052  0.2143 -0.7024  0.3346 -0.2376  0.7453 -0.2973
#>   0.3448 -0.7037 -0.2405  0.2290  0.4304  0.3876  0.0109  0.2700 -0.2987
#> 
#> Columns 19 to 20 -0.2159  0.3868
#>  -0.1437 -0.1137
#>   0.3752 -0.1488
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5998 -0.2870  0.2525  0.0801 -0.1137  0.0290 -0.0548 -0.4434  0.0166
#>  -0.5544  0.4798  0.3190 -0.4991  0.0824 -0.5724  0.0300 -0.2911 -0.1159
#>  -0.3525  0.2093  0.4062 -0.4659 -0.3296 -0.3556 -0.2793 -0.3885 -0.1262
#> 
#> Columns 10 to 18  0.4041 -0.3267  0.1758 -0.0003  0.4376  0.2877 -0.2924  0.1986  0.1473
#>   0.5377 -0.1934  0.2942 -0.0647 -0.1449  0.0177 -0.6033  0.5790 -0.2270
#>   0.4109 -0.2481  0.1317  0.0488 -0.0593 -0.3978 -0.5629  0.3926 -0.1205
#> 
#> Columns 19 to 20  0.2438  0.0659
#>   0.4845  0.1527
#>   0.1884  0.3544
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>