RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3222 -0.2191  0.8306  0.0711 -0.7519  0.1613  0.0940 -0.7961  0.8526
#>   0.5483 -0.5197  0.7647  0.1072 -0.5048 -0.1786 -0.6646 -0.1379 -0.2405
#>  -0.4826 -0.4066 -0.2892  0.5713 -0.8559  0.8629  0.4672  0.0668 -0.4364
#> 
#> Columns 10 to 18 -0.3458  0.6421 -0.3783  0.1220  0.6044  0.1721 -0.0413  0.4672 -0.7573
#>  -0.5987  0.0229  0.2509  0.1869  0.4065  0.0932  0.1174 -0.3268  0.6951
#>   0.1783 -0.8110 -0.0930 -0.2823  0.6164  0.6293 -0.0176  0.7341 -0.4645
#> 
#> Columns 19 to 20 -0.7456  0.2742
#>  -0.2046 -0.7739
#>   0.4720 -0.1913
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4643  0.0800  0.1624  0.1012 -0.7024  0.5377 -0.2109  0.0399 -0.6164
#>   0.3538 -0.1074 -0.3341  0.5066  0.0360  0.5416  0.1964 -0.0746  0.3112
#>   0.2209 -0.7143  0.1834  0.3932  0.0460  0.0608 -0.5590 -0.2415 -0.1361
#> 
#> Columns 10 to 18  0.0137  0.5565 -0.4432  0.2450  0.6199  0.4625  0.4132  0.0553 -0.7390
#>   0.1090 -0.0517 -0.5911 -0.2543  0.1513  0.0212  0.3803 -0.3625 -0.5126
#>  -0.0432 -0.0712  0.2833 -0.0696  0.6083  0.2226  0.3615  0.1737  0.0535
#> 
#> Columns 19 to 20 -0.0166  0.5152
#>  -0.3413  0.6821
#>  -0.4681 -0.4582
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1675 -0.6907  0.2123  0.1298 -0.6279  0.7304  0.2911  0.3961 -0.2490
#>   0.2748 -0.3973  0.3963 -0.0392 -0.6652  0.4636 -0.1311 -0.5021 -0.3061
#>   0.6042 -0.2195 -0.0388  0.4817 -0.3584  0.7711 -0.2759  0.2799 -0.3350
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2236 -0.3340  0.8450 -0.5684  0.7371  0.2016  0.0625  0.4266 -0.1081
#>  -0.1240  0.3061  0.1657 -0.2923  0.4016 -0.3768  0.6406 -0.1284  0.1972
#>  -0.5896 -0.2507 -0.5967 -0.5883 -0.0749 -0.0087 -0.2863  0.1277  0.5248
#> 
#> Columns 10 to 18  0.3230 -0.3561 -0.0823  0.4184 -0.5393  0.6555  0.5140 -0.5120 -0.4632
#>  -0.0976 -0.6183  0.2294 -0.6237  0.1033  0.4195  0.4203 -0.1337 -0.3996
#>   0.3206  0.8829  0.0637 -0.3422 -0.0586 -0.0387  0.7329  0.5338  0.6542
#> 
#> Columns 19 to 20 -0.5455 -0.2512
#>  -0.5378  0.6921
#>   0.2714  0.3554
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0413 -0.4854  0.1407  0.2398 -0.3665  0.4430  0.1905 -0.3957  0.2053
#>  -0.2649 -0.4229  0.2133 -0.0305 -0.5382 -0.0173  0.0362 -0.1686  0.0262
#>   0.4767 -0.3636  0.2162 -0.0220 -0.5902  0.7084 -0.4412  0.2063 -0.0020
#> 
#> Columns 10 to 18  0.2016  0.3056  0.0368  0.1267  0.7467  0.1783  0.0964  0.6932 -0.3230
#>   0.1747  0.7229  0.0363 -0.3089  0.5950 -0.4342 -0.0568  0.2915  0.0646
#>  -0.3690 -0.0536  0.1282  0.4543  0.2436 -0.0892  0.3764 -0.0453 -0.6057
#> 
#> Columns 19 to 20 -0.6946  0.3768
#>  -0.4000 -0.0043
#>  -0.4477  0.4121
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>