RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0410  0.0815 -0.6317 -0.7146 -0.1833  0.0877 -0.4484  0.1138 -0.3400
#>   0.3172  0.1347 -0.7118  0.0442  0.4597  0.1109 -0.7985  0.9620  0.1666
#>   0.8267 -0.0176  0.3935  0.4470 -0.3475 -0.1524  0.6240  0.2422 -0.4040
#> 
#> Columns 10 to 18 -0.4812 -0.6224  0.1446  0.5502 -0.0279 -0.2613 -0.4872  0.5829 -0.7423
#>   0.0802 -0.2563 -0.8969  0.6552 -0.1728 -0.7266  0.2526 -0.1946 -0.5535
#>   0.3104  0.8730  0.4011  0.0846 -0.4718  0.1780  0.3278  0.2170  0.9666
#> 
#> Columns 19 to 20  0.5615  0.7298
#>  -0.9097  0.9173
#>  -0.1355  0.7415
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6645  0.2635 -0.2840  0.5293  0.4920  0.1607 -0.3985 -0.2887  0.6356
#>   0.1172  0.7432 -0.7662  0.5250 -0.2047  0.3950 -0.5903 -0.0988  0.6720
#>   0.4415  0.2351 -0.0760 -0.0815  0.4962 -0.2087 -0.3593 -0.1552  0.1967
#> 
#> Columns 10 to 18 -0.5177 -0.0069 -0.7664  0.4865 -0.1757 -0.5942  0.5632  0.0149  0.4605
#>  -0.2768  0.1877 -0.8613  0.5188 -0.5399 -0.3450  0.5787 -0.0556  0.4297
#>  -0.3978  0.2410  0.1343 -0.2447 -0.0461 -0.4259  0.4429  0.4072  0.2790
#> 
#> Columns 19 to 20 -0.2956  0.4280
#>  -0.3888  0.4469
#>  -0.0627  0.2266
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5617  0.1655 -0.5787  0.4210 -0.1085 -0.2678  0.5173 -0.5048  0.2965
#>  -0.2507  0.1941 -0.7930  0.0815  0.0396 -0.1575  0.1455 -0.2856  0.4389
#>   0.4635  0.2417 -0.3616  0.0999  0.0830  0.0298  0.0056  0.0726  0.3404
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0322  0.1385  0.1011  0.1129 -0.0104  0.5389  0.1661  0.8130 -0.1338
#>  -0.1288 -0.1987 -0.3632  0.3414 -0.4970  0.2313 -0.5062  0.3551  0.1196
#>  -0.2525 -0.3297 -0.4547  0.6389 -0.4539 -0.6702 -0.4087 -0.1213 -0.2057
#> 
#> Columns 10 to 18 -0.1874  0.6539 -0.3435  0.2091  0.0554  0.2135  0.5233  0.2263 -0.0734
#>  -0.1007 -0.3277 -0.0182 -0.3177 -0.7331 -0.1205  0.3400  0.3809 -0.1523
#>   0.2309 -0.2493  0.1633 -0.2352 -0.5113 -0.2193 -0.0021  0.0800 -0.5350
#> 
#> Columns 19 to 20  0.2235 -0.0371
#>   0.1104  0.5801
#>   0.8019  0.1268
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2377  0.2140 -0.5241  0.3905  0.5335  0.0369 -0.0270  0.1011  0.3565
#>  -0.1730  0.0727 -0.6696 -0.1398 -0.0077 -0.1154  0.3137 -0.1466 -0.1065
#>  -0.2612 -0.1942 -0.5651  0.3101  0.0293 -0.3832  0.5577  0.3558 -0.1813
#> 
#> Columns 10 to 18 -0.1308 -0.0524 -0.4059  0.3237 -0.0726 -0.3336  0.3226  0.3731  0.2451
#>  -0.2336  0.3014 -0.2902  0.3210 -0.4916  0.3277  0.0126  0.0013  0.1941
#>  -0.3538  0.2530  0.4592  0.1638 -0.4497 -0.2407 -0.2059  0.0581  0.1766
#> 
#> Columns 19 to 20 -0.3438  0.3165
#>  -0.5664  0.2581
#>  -0.0850 -0.2782
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>