RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2943  0.7088 -0.7771  0.2536  0.2226  0.6205 -0.2690 -0.2238 -0.9117
#>   0.5068 -0.2346  0.5276  0.6645  0.5879 -0.9464  0.0201 -0.2742  0.0848
#>   0.8190 -0.2323  0.5764 -0.2375 -0.3340  0.6155 -0.2395  0.5698 -0.1896
#> 
#> Columns 10 to 18  0.1600  0.8359  0.0188  0.0490 -0.3135 -0.5712 -0.5880 -0.6124  0.6206
#>   0.2201 -0.3566  0.2951 -0.6497  0.8281  0.2820 -0.6385  0.4676  0.3842
#>   0.3547 -0.7471 -0.0662 -0.7447  0.3171 -0.6878  0.4679 -0.5434  0.3796
#> 
#> Columns 19 to 20  0.7486 -0.5350
#>   0.7655  0.5784
#>   0.6723  0.7869
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5096  0.3392  0.6660  0.1756  0.0488  0.0322  0.0872 -0.5956  0.4919
#>  -0.3347 -0.0684  0.2870 -0.2096  0.2778  0.0196 -0.0188 -0.5749  0.2614
#>   0.0572  0.3414  0.4550  0.3972  0.3049 -0.3675 -0.3509 -0.4664 -0.8277
#> 
#> Columns 10 to 18 -0.1018 -0.1095  0.4218  0.5094  0.3999 -0.2756  0.8046  0.8085 -0.1439
#>  -0.2331 -0.8364 -0.2636 -0.4478  0.2578  0.7367 -0.1090 -0.2337  0.2447
#>   0.2291 -0.2044  0.2358 -0.6434 -0.3135 -0.3221  0.2179  0.0605  0.1244
#> 
#> Columns 19 to 20  0.1706  0.2426
#>  -0.1359  0.1262
#>   0.0418  0.0957
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0162  0.4418 -0.5645  0.1696  0.2147 -0.2683 -0.1682 -0.1994 -0.3600
#>  -0.5257  0.3450 -0.0432 -0.0664  0.5317  0.0696  0.1445 -0.5587  0.2429
#>  -0.4874  0.6947 -0.0582  0.5251  0.1589  0.0299 -0.0509 -0.5755 -0.1996
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2438 -0.7503 -0.0013 -0.3172 -0.5996 -0.1455 -0.5211 -0.5346  0.4405
#>   0.1568 -0.2878 -0.1508 -0.3451 -0.4863  0.2569 -0.1561 -0.7374  0.8137
#>   0.4188  0.1050 -0.3095  0.5409 -0.5596 -0.1709 -0.0904  0.5327 -0.1670
#> 
#> Columns 10 to 18  0.0516 -0.0289  0.0846 -0.6013  0.4206 -0.6188  0.2063  0.4905 -0.1980
#>  -0.6958  0.0544  0.3639  0.0937 -0.2943 -0.1105 -0.0065  0.1094 -0.8953
#>  -0.1116  0.0775  0.1432 -0.3213 -0.4223  0.3944 -0.2291  0.0527 -0.5023
#> 
#> Columns 19 to 20 -0.6151  0.6305
#>   0.3117  0.3178
#>   0.6196  0.1933
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3081  0.4119 -0.1076  0.2220  0.5091  0.1863  0.0819 -0.3111  0.3973
#>  -0.3370  0.2352 -0.1377  0.2262  0.1964  0.2842  0.0540 -0.6465  0.4908
#>   0.1541  0.1588 -0.3036  0.0169  0.1859 -0.0433 -0.2080  0.0752 -0.3214
#> 
#> Columns 10 to 18  0.0873 -0.4637  0.3160  0.0583  0.0385 -0.2930  0.4585  0.1767  0.2024
#>  -0.1781 -0.4963  0.6029 -0.1773  0.2743 -0.1597  0.2081  0.2906  0.2617
#>   0.2088 -0.2366  0.4525 -0.1572  0.3944 -0.2015 -0.4393  0.2490  0.0460
#> 
#> Columns 19 to 20  0.2091  0.3098
#>   0.4717  0.0439
#>   0.5071  0.4477
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>