RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6373 -0.7800  0.5391 -0.4817 -0.1869  0.0638  0.7749  0.5801 -0.3204
#>   0.9168 -0.5574 -0.6866  0.3149 -0.4315 -0.6744 -0.8379  0.6298 -0.5897
#>   0.8045  0.6545  0.4364  0.7698 -0.7138  0.5105 -0.5996 -0.0532 -0.5612
#> 
#> Columns 10 to 18 -0.3285 -0.2932 -0.0675  0.3761  0.3227  0.7124 -0.8032  0.8753  0.9165
#>   0.3460  0.2881  0.1059  0.0527 -0.3421  0.6926 -0.4507 -0.1793  0.1230
#>  -0.6848 -0.8058 -0.1401 -0.4799 -0.7227 -0.3970  0.7250  0.2161 -0.7719
#> 
#> Columns 19 to 20 -0.2347  0.4776
#>  -0.3076 -0.0283
#>   0.5415  0.5301
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1093 -0.2509 -0.3816 -0.1517 -0.4801  0.4107  0.3801  0.3815 -0.4712
#>  -0.1930  0.1213 -0.1176 -0.6126  0.4643  0.2591  0.1384  0.4842  0.4212
#>   0.6036 -0.7561 -0.6294  0.2085  0.1144 -0.3158  0.1732  0.0168 -0.4493
#> 
#> Columns 10 to 18 -0.0188 -0.7082 -0.3932 -0.8705 -0.3027 -0.4170 -0.2315  0.3913 -0.3412
#>  -0.3101 -0.0199 -0.2291  0.2940 -0.6417  0.6073 -0.1995  0.7858  0.3032
#>  -0.7247  0.5028  0.1431  0.3633 -0.2370  0.3758 -0.2093 -0.3286  0.1501
#> 
#> Columns 19 to 20 -0.1439  0.3706
#>  -0.6097  0.0930
#>  -0.7308  0.2277
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4642 -0.7529  0.1196  0.2340 -0.2620  0.0897  0.2084  0.2181 -0.1635
#>   0.2734 -0.5309 -0.3936 -0.1842 -0.2136  0.4762  0.4240  0.4648  0.1212
#>  -0.2062  0.1346  0.5883 -0.0879  0.0285  0.1545 -0.0914  0.5267  0.3968
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6774 -0.3834 -0.5537  0.4762  0.5122 -0.0255 -0.1872 -0.5326 -0.3103
#>   0.7259  0.2183  0.3075 -0.2676 -0.4325  0.7134  0.2669  0.5812  0.9142
#>   0.7405 -0.5972  0.0302  0.1597 -0.4158 -0.0195 -0.2261  0.5193  0.3749
#> 
#> Columns 10 to 18 -0.2627 -0.7954  0.2499 -0.2880 -0.4395 -0.0877  0.3200 -0.1359 -0.1963
#>   0.4452  0.5040 -0.3778  0.3074 -0.0337  0.2087 -0.8191  0.5803  0.3159
#>   0.7186  0.0024  0.2310 -0.3444  0.0854  0.3467 -0.2982 -0.1140  0.1253
#> 
#> Columns 19 to 20 -0.6423  0.2325
#>   0.7141  0.5251
#>  -0.5787 -0.1369
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2876 -0.4595  0.2439 -0.3154  0.2983  0.1112  0.4227  0.2631 -0.5486
#>  -0.0771 -0.6936  0.1298  0.1592 -0.2769 -0.1928  0.1926  0.3570 -0.4857
#>  -0.1366 -0.7209  0.1294  0.1070 -0.3476  0.3612  0.1587 -0.0804  0.1929
#> 
#> Columns 10 to 18  0.2410 -0.4209  0.2953 -0.1544 -0.4258  0.6672 -0.0284  0.5949  0.4125
#>   0.2098 -0.5466 -0.2345 -0.3216 -0.1359  0.1133  0.2491  0.5593 -0.3741
#>  -0.0544  0.0492 -0.0059 -0.5532  0.0380  0.2602 -0.0058  0.7465 -0.0982
#> 
#> Columns 19 to 20 -0.3915 -0.2400
#>  -0.3666  0.3623
#>  -0.0937  0.2179
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>