RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6459 -0.7705 -0.5107 -0.6023  0.3283 -0.5689 -0.7770  0.8414  0.2036
#>   0.7247  0.0392  0.5309  0.5374  0.7565  0.2029 -0.8682  0.1162  0.7755
#>   0.8195 -0.1588  0.6454 -0.7435  0.0161  0.0091 -0.5944 -0.0739  0.7689
#> 
#> Columns 10 to 18 -0.2536 -0.1230  0.2172  0.3383  0.6208 -0.1054  0.0902  0.4940 -0.7418
#>   0.4498 -0.8450  0.3896  0.2231 -0.8754 -0.0007  0.2779 -0.4304 -0.7161
#>   0.0197  0.2848  0.3321  0.7642  0.6028 -0.5746 -0.1989  0.5597 -0.3842
#> 
#> Columns 19 to 20 -0.9332  0.5420
#>  -0.3889 -0.2376
#>   0.0002 -0.2864
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0186  0.3364 -0.1248 -0.2443 -0.0729 -0.1663  0.2442  0.4029  0.4262
#>   0.0139 -0.1046  0.2661 -0.4345  0.3308 -0.1485 -0.1541  0.3100  0.5633
#>  -0.2983 -0.6215  0.1647 -0.5566  0.1211  0.2958 -0.4943 -0.2098 -0.2869
#> 
#> Columns 10 to 18 -0.2450 -0.2471  0.3062 -0.0761 -0.6203  0.2362 -0.4673  0.7161 -0.8082
#>  -0.0342  0.1970 -0.2313  0.3291 -0.2341 -0.4768 -0.5172  0.0355 -0.4182
#>   0.1856  0.2374 -0.4336  0.3772 -0.3164 -0.1149 -0.1996  0.2703  0.1483
#> 
#> Columns 19 to 20  0.2304 -0.4304
#>  -0.1874  0.2656
#>  -0.2614 -0.0367
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1941 -0.4210  0.4829 -0.3581 -0.0310  0.2282 -0.3170  0.0146  0.5160
#>   0.3671  0.0984  0.4894 -0.4071 -0.3597  0.1830 -0.4094  0.0905  0.2904
#>  -0.0659 -0.3747  0.3155 -0.4730  0.0124 -0.0905 -0.2564 -0.1667  0.2675
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0085 -0.0447  0.0176 -0.3704 -0.1104 -0.7658 -0.4517 -0.0363 -0.4589
#>   0.3953  0.4256 -0.0372  0.4473  0.8030  0.0747 -0.6207 -0.6180 -0.7457
#>   0.4599  0.1010  0.8767  0.4409  0.8999  0.4533  0.0891  0.8494 -0.2091
#> 
#> Columns 10 to 18 -0.7096 -0.1350  0.2392  0.5018  0.0639 -0.4980  0.2689  0.7355  0.5697
#>   0.1068  0.3819 -0.6024  0.0946 -0.0375  0.1427 -0.2807  0.3150 -0.3451
#>   0.0506  0.1718 -0.7208  0.2593  0.0792  0.0562  0.0732 -0.6050  0.4799
#> 
#> Columns 19 to 20 -0.5767  0.4801
#>   0.1015  0.3044
#>  -0.0440  0.5650
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2495 -0.4331  0.3936 -0.3148 -0.2010  0.1748 -0.4654 -0.0877  0.1114
#>  -0.2534 -0.5892  0.5209 -0.4487  0.1162  0.0552 -0.4628 -0.2052  0.4286
#>   0.3390 -0.1804  0.4130 -0.5892  0.2807  0.0290 -0.2145 -0.3031  0.7140
#> 
#> Columns 10 to 18  0.5967 -0.4008 -0.0048  0.5541 -0.0006 -0.4303 -0.2266  0.3871 -0.2341
#>  -0.3677  0.3574 -0.1504  0.7231 -0.4926 -0.3051 -0.6201  0.3274  0.0490
#>   0.0983  0.3406 -0.3145  0.0773 -0.1610 -0.4193 -0.2214  0.4596 -0.2660
#> 
#> Columns 19 to 20  0.1026 -0.2677
#>  -0.6480  0.4080
#>  -0.2600 -0.1124
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>