RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2627  0.6525 -0.9705 -0.5686  0.3755  0.2596 -0.0081 -0.8348  0.6246
#>  -0.6322  0.8298 -0.2267 -0.6232  0.6293 -0.2709 -0.6448  0.0468  0.3510
#>  -0.0234 -0.7283 -0.4513 -0.7208  0.6752  0.0285 -0.5918 -0.1879 -0.6594
#> 
#> Columns 10 to 18 -0.3971  0.2263 -0.4535  0.0837  0.8349 -0.6836  0.8424 -0.7597  0.4923
#>   0.5435  0.5312 -0.5934 -0.6723  0.8188  0.0018 -0.2492  0.9448  0.3724
#>   0.1548 -0.5359 -0.4880 -0.7727  0.0066  0.8179 -0.4918  0.7017 -0.4879
#> 
#> Columns 19 to 20  0.1808 -0.0387
#>   0.7501 -0.5928
#>   0.3459 -0.7171
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4247 -0.3077 -0.4550  0.3576 -0.0451 -0.3646  0.2589  0.1491 -0.4797
#>   0.1449 -0.3928  0.1453  0.0591 -0.5251 -0.4176 -0.5549  0.0801 -0.5353
#>  -0.2621  0.5993 -0.3112 -0.5093 -0.3938 -0.0609 -0.5198 -0.4562  0.0251
#> 
#> Columns 10 to 18 -0.0726 -0.5831 -0.3839 -0.3642 -0.2549  0.1816 -0.2277  0.3741  0.1478
#>  -0.3935 -0.3793 -0.4468  0.3163 -0.4515  0.5245  0.4168 -0.0156 -0.2378
#>   0.1455 -0.3239 -0.3497 -0.2507  0.1654  0.3706  0.3980  0.4592  0.5174
#> 
#> Columns 19 to 20 -0.3370 -0.1030
#>  -0.4196 -0.3438
#>  -0.0792 -0.5128
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0045  0.1653 -0.3249  0.3177 -0.4288 -0.1408 -0.4598 -0.0957 -0.1727
#>  -0.2745  0.4476 -0.4309 -0.1344  0.1033  0.1228 -0.0211  0.2102 -0.0013
#>   0.2226  0.0166 -0.2129 -0.0754  0.0441 -0.4264 -0.0234 -0.3483  0.0884
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0840  0.7141  0.7086  0.2821 -0.2528 -0.1725  0.5318 -0.2844 -0.3398
#>  -0.0485  0.6878  0.6344  0.4824  0.6335  0.4477 -0.2199  0.6811 -0.2497
#>  -0.5205  0.3073  0.4490 -0.1455  0.2095 -0.3496 -0.0474 -0.6100  0.0046
#> 
#> Columns 10 to 18 -0.3942  0.1930  0.5002  0.7075  0.4311  0.1580 -0.5208  0.1056 -0.2842
#>  -0.7051  0.2181  0.2323  0.8433 -0.2469  0.6341 -0.7362  0.7898  0.3570
#>  -0.5385  0.3342  0.2458  0.8319 -0.3787 -0.4605 -0.4756 -0.6434  0.0812
#> 
#> Columns 19 to 20 -0.6125  0.0851
#>  -0.4995  0.1148
#>  -0.8209 -0.1639
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1892  0.0203 -0.3884  0.1575 -0.4549 -0.2956 -0.6065 -0.2504 -0.0073
#>   0.3725 -0.2907 -0.2578  0.1676 -0.1994  0.3646 -0.5102 -0.3160  0.3151
#>   0.0586  0.2978 -0.5652 -0.0082  0.1840 -0.1416 -0.1189 -0.0486 -0.0827
#> 
#> Columns 10 to 18 -0.4863 -0.5838 -0.5912 -0.0888 -0.3452  0.4245  0.6024  0.1342  0.1875
#>  -0.3267  0.0067 -0.7477 -0.3690  0.3413 -0.2749  0.8188  0.0906  0.4216
#>   0.0081 -0.5507 -0.6006 -0.5923 -0.0260 -0.0574  0.6906  0.1161  0.5917
#> 
#> Columns 19 to 20 -0.5233 -0.1297
#>  -0.5040 -0.3348
#>  -0.5062 -0.5686
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>