RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6568 -0.0494  0.2016  0.2568 -0.4566  0.0430  0.5853  0.7720 -0.6809
#>  -0.2471 -0.7801  0.5921  0.1339  0.4882 -0.8853 -0.0819  0.3202 -0.7137
#>  -0.0437 -0.8866  0.2328  0.9693  0.4998 -0.9085 -0.6758 -0.0370 -0.0654
#> 
#> Columns 10 to 18 -0.8084 -0.5101 -0.5385 -0.5564  0.1339 -0.4565 -0.0828  0.8337 -0.8335
#>  -0.3559 -0.5487  0.8769 -0.6092  0.9253  0.7728  0.6246  0.4038 -0.4464
#>  -0.4467  0.5392  0.6582 -0.8940  0.8677  0.3858 -0.2421  0.5594 -0.9404
#> 
#> Columns 19 to 20 -0.2753  0.2950
#>   0.6432 -0.3685
#>   0.5896 -0.6819
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1222 -0.2986  0.2652  0.6073  0.2139 -0.1937 -0.5310  0.0229 -0.4246
#>  -0.3553 -0.8607  0.4486  0.6856  0.0923 -0.3683 -0.2492  0.0894 -0.3629
#>   0.0839 -0.6184  0.4692  0.0017 -0.0666 -0.1115 -0.2058  0.1220 -0.0431
#> 
#> Columns 10 to 18  0.3029 -0.5266 -0.5771 -0.0941 -0.1493 -0.1016  0.0398 -0.1365 -0.4112
#>   0.2528 -0.0683 -0.0366  0.2812  0.1991  0.0173  0.3952  0.4745 -0.1190
#>   0.0128 -0.3323 -0.2525 -0.5337  0.1939  0.4098  0.6538  0.4603 -0.5788
#> 
#> Columns 19 to 20 -0.2229  0.2434
#>  -0.5748  0.0041
#>  -0.0367 -0.2261
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3547 -0.4695  0.4114  0.5944  0.1503 -0.2670  0.3646  0.3322  0.1516
#>  -0.4033 -0.7917  0.3639  0.3355  0.3718  0.2410  0.1933 -0.1048 -0.1602
#>  -0.3049 -0.5926  0.2576  0.5497  0.1654 -0.1376 -0.3503 -0.2201 -0.2103
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5045 -0.7022  0.5228  0.1349  0.4436 -0.3656  0.3504  0.7999 -0.1158
#>   0.2820 -0.1929  0.5658 -0.0534  0.5758  0.1499  0.0144 -0.7372  0.5040
#>   0.7623 -0.2300 -0.5686 -0.2656  0.2368  0.5234 -0.0535  0.5754  0.5739
#> 
#> Columns 10 to 18  0.0311 -0.2649  0.0809  0.3690  0.1976 -0.4013  0.2549 -0.5611  0.1666
#>   0.1850  0.4734 -0.6014  0.1997 -0.1802  0.8172  0.6496 -0.4878 -0.2766
#>  -0.7921 -0.4462  0.6168  0.0923 -0.0545  0.7265 -0.6057 -0.7283  0.0272
#> 
#> Columns 19 to 20 -0.5575 -0.2823
#>   0.8196 -0.2373
#>   0.5103 -0.6279
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4610 -0.6359  0.5009  0.0488  0.6907  0.3112  0.1418  0.0058 -0.2569
#>  -0.5095 -0.5345  0.2455  0.8495  0.0612 -0.1179 -0.0802  0.1494 -0.5245
#>  -0.1737 -0.4484  0.1765  0.0972  0.0746 -0.1412 -0.0005 -0.6862  0.3568
#> 
#> Columns 10 to 18  0.3227 -0.3523 -0.1879 -0.1552 -0.0660  0.1961  0.0374  0.1047 -0.2251
#>   0.3528  0.1014 -0.4005  0.6024  0.0379 -0.3944  0.4516  0.2555  0.1841
#>   0.3455  0.0278 -0.7181 -0.2459  0.0668  0.0290  0.2264  0.1190 -0.7264
#> 
#> Columns 19 to 20 -0.0257 -0.5122
#>  -0.1046 -0.1982
#>   0.4000 -0.3852
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>