RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2738  0.1272  0.4977 -0.4503  0.6442 -0.0028  0.6426  0.5765 -0.4748
#>  -0.4664  0.9122 -0.7631 -0.5587  0.5439  0.4066 -0.6715  0.4675  0.5892
#>   0.2068 -0.0165 -0.2690 -0.4673  0.4174  0.2049  0.7366  0.1616 -0.7172
#> 
#> Columns 10 to 18  0.5408 -0.7151  0.3636  0.9072  0.6374  0.3778 -0.1920 -0.2061  0.4690
#>  -0.4958 -0.4524 -0.4439  0.5462 -0.7070 -0.0975 -0.3767  0.4875 -0.1465
#>  -0.2516 -0.4415 -0.4915 -0.3549  0.7055 -0.1536  0.5833 -0.5385  0.2087
#> 
#> Columns 19 to 20 -0.3475 -0.1348
#>   0.6293  0.0892
#>   0.7026 -0.3894
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1835 -0.2719  0.2769  0.5393  0.5268  0.4308 -0.0079  0.5935  0.6299
#>   0.2103  0.0774 -0.5905 -0.1555  0.1198  0.2429  0.2271 -0.5022  0.5001
#>  -0.2446  0.5222 -0.2316 -0.2547 -0.0686 -0.2297  0.2083 -0.1389 -0.4555
#> 
#> Columns 10 to 18 -0.3433 -0.7022 -0.3525  0.6354  0.0509  0.1582 -0.2591  0.3339  0.4536
#>  -0.4758  0.3214 -0.0500 -0.5912 -0.1408  0.3667 -0.5127 -0.1789 -0.6625
#>   0.3817 -0.2821  0.0399 -0.0361  0.1439  0.4725 -0.1357 -0.0894 -0.1539
#> 
#> Columns 19 to 20 -0.0237  0.0744
#>   0.4044 -0.1660
#>  -0.1783 -0.4996
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1475  0.2283 -0.0015 -0.1988 -0.2166  0.0678  0.5320 -0.6607 -0.4714
#>  -0.4710  0.5106 -0.2081 -0.6182 -0.1809  0.3342  0.2716 -0.2104  0.0388
#>  -0.3317  0.3230  0.0760 -0.1373  0.2893  0.2274  0.0205 -0.1013  0.2599
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5040 -0.0986 -0.3351  0.0103  0.4008  0.5927 -0.2572  0.7046  0.1302
#>   0.2888  0.1532  0.2120  0.5225  0.2752  0.2578  0.5204  0.4539  0.8976
#>   0.2663  0.3757  0.6803 -0.3552 -0.0155  0.3584  0.6981  0.4924  0.3806
#> 
#> Columns 10 to 18  0.0567  0.5685  0.1307 -0.2816 -0.0795  0.2806 -0.1880  0.3690  0.5232
#>   0.0471  0.2623  0.0590  0.5976  0.5394 -0.5880  0.2497 -0.2903  0.3251
#>  -0.1618 -0.4424 -0.0514 -0.4424 -0.0584  0.2698 -0.1784  0.0250 -0.6076
#> 
#> Columns 19 to 20 -0.4477 -0.5699
#>   0.6487 -0.8117
#>  -0.2457 -0.3905
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1063 -0.0336  0.3422  0.2592 -0.0034  0.5387  0.3054 -0.2262 -0.0349
#>  -0.1745  0.0001 -0.0367 -0.4387 -0.4229 -0.0848  0.2199 -0.4320 -0.2791
#>  -0.3640 -0.0314  0.4414 -0.2267 -0.4257  0.3334  0.2641 -0.2300  0.5366
#> 
#> Columns 10 to 18 -0.2173 -0.6254  0.4287  0.3236 -0.4317  0.6303 -0.4590 -0.4281 -0.1787
#>  -0.0688 -0.0892 -0.1975 -0.2371 -0.2353  0.7905 -0.0764 -0.1019 -0.2407
#>  -0.5394  0.0206  0.2609  0.0150 -0.5179  0.4849  0.0323  0.0060  0.1051
#> 
#> Columns 19 to 20 -0.2577  0.2544
#>   0.1758 -0.5506
#>   0.2800 -0.1873
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>