RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3125 -0.5230  0.7288 -0.7173 -0.7147  0.8544  0.3586  0.6362 -0.0468
#>   0.2827  0.0585  0.3934  0.3996 -0.5573  0.1533  0.4959  0.7036 -0.6124
#>   0.0843 -0.8247  0.2136  0.3747  0.3005 -0.6798 -0.4259 -0.2485  0.0993
#> 
#> Columns 10 to 18 -0.7433  0.0458 -0.0714 -0.3878  0.8276 -0.3007  0.8881  0.5619  0.0323
#>   0.2131  0.5593 -0.1992 -0.4913  0.4031  0.5414 -0.0392  0.7267  0.6858
#>  -0.2784  0.5957  0.4345 -0.8714  0.7372 -0.1533 -0.7308  0.7475 -0.7670
#> 
#> Columns 19 to 20 -0.7813 -0.3236
#>  -0.2293  0.9228
#>   0.7347  0.1313
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1924 -0.1198  0.2073 -0.0345 -0.2051 -0.0451 -0.0299  0.0347  0.1752
#>  -0.1171 -0.2332  0.3344 -0.3049  0.4429 -0.1594  0.4329 -0.3213  0.0818
#>   0.1807 -0.5543  0.3937 -0.3186 -0.3472 -0.3799 -0.0668 -0.0493 -0.0816
#> 
#> Columns 10 to 18 -0.4300 -0.5001  0.5920 -0.1868 -0.2926  0.1346  0.3736  0.4944  0.6150
#>  -0.1445 -0.0818  0.2136 -0.1240  0.1152 -0.4528 -0.1838 -0.1173 -0.1857
#>  -0.1868  0.1400 -0.5466 -0.0838  0.3798 -0.3395  0.1115  0.0602 -0.3393
#> 
#> Columns 19 to 20  0.3812  0.5355
#>   0.3386 -0.3030
#>   0.4535 -0.5035
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3830 -0.0952  0.2322  0.0153 -0.0560 -0.4110  0.3976  0.1133  0.1572
#>   0.0769 -0.0110  0.5244  0.1222 -0.1137 -0.2058 -0.5852 -0.0713 -0.1567
#>  -0.2549  0.2352  0.4812 -0.2285  0.3170  0.2980 -0.2952  0.3851 -0.1892
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1849 -0.6642 -0.6812 -0.6210 -0.2958 -0.4492 -0.7081  0.4074  0.7183
#>   0.0413 -0.3254 -0.7071 -0.7961  0.1535  0.1923 -0.8095  0.4486  0.6496
#>   0.9117 -0.0351  0.1031  0.1085 -0.2764  0.2688 -0.0928 -0.0467  0.7945
#> 
#> Columns 10 to 18  0.0365 -0.0326 -0.5778  0.3037 -0.4763  0.2992  0.3360 -0.4228  0.4502
#>  -0.5620 -0.2400 -0.0388 -0.1218 -0.5622 -0.2689  0.1808 -0.3312  0.1295
#>  -0.8934 -0.6643 -0.2485  0.2837  0.2489  0.4316 -0.7444  0.5356  0.7766
#> 
#> Columns 19 to 20  0.1529 -0.1264
#>  -0.5166 -0.1125
#>  -0.6670  0.7650
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1763  0.3803  0.0803 -0.2615  0.1524 -0.2713 -0.4083 -0.1161  0.1286
#>   0.0766  0.1559  0.3014 -0.2320  0.4172 -0.0095  0.0246  0.1345 -0.0116
#>   0.0929  0.1528  0.4656 -0.0123 -0.0898 -0.4063  0.5038  0.1751 -0.3089
#> 
#> Columns 10 to 18  0.3953 -0.0101  0.1363 -0.2940  0.2283  0.1231 -0.4609  0.6638 -0.2285
#>   0.3573 -0.0973  0.1334 -0.3110  0.1742 -0.1298 -0.1582  0.6952 -0.4107
#>   0.4001  0.5398 -0.1454 -0.1928  0.0954 -0.3087 -0.0543  0.6525 -0.7035
#> 
#> Columns 19 to 20 -0.3998 -0.0903
#>  -0.3388 -0.0500
#>   0.5733 -0.1503
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>