RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6875 -0.1081  0.8098 -0.7260 -0.7253  0.3970  0.0270 -0.6974 -0.3162
#>   0.2054 -0.6937  0.2555  0.5556  0.8656  0.1725  0.3952 -0.4965 -0.2291
#>  -0.3184 -0.3346  0.5733  0.4788  0.8926  0.0034  0.3443 -0.0263  0.6963
#> 
#> Columns 10 to 18  0.5736  0.0880  0.5676 -0.6106  0.5787 -0.5265  0.0590 -0.0011 -0.4850
#>   0.1197 -0.3779  0.5613 -0.2148 -0.7042  0.2840 -0.4051 -0.6803 -0.8451
#>   0.1723 -0.2076  0.4839 -0.2177  0.0404  0.7889 -0.2866 -0.7888 -0.4214
#> 
#> Columns 19 to 20  0.2023 -0.3177
#>  -0.5455  0.5691
#>  -0.4437  0.6308
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0761 -0.1368  0.1638  0.1301 -0.3848 -0.2311  0.1354  0.0772 -0.2278
#>  -0.4870  0.4847  0.0199  0.2579  0.4644 -0.4233 -0.5670 -0.1848 -0.5921
#>  -0.0106 -0.0811  0.6294  0.0686  0.6251 -0.3487 -0.0917 -0.0863 -0.2318
#> 
#> Columns 10 to 18  0.1718 -0.2388 -0.4193 -0.1432 -0.3720  0.3100  0.3099 -0.3395  0.0474
#>  -0.0455  0.2167  0.5574 -0.1362 -0.3924  0.0116  0.3409 -0.6544  0.3165
#>   0.3223  0.2149  0.1528 -0.5181  0.3834  0.1948 -0.4920 -0.4009 -0.1649
#> 
#> Columns 19 to 20 -0.3096  0.5888
#>   0.6266  0.5493
#>   0.0480  0.2675
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0174  0.1336  0.1845  0.4566  0.5153  0.0823  0.3840 -0.5575 -0.2606
#>  -0.2055 -0.1859  0.0131  0.5298  0.5459 -0.1679  0.0693  0.4961 -0.3962
#>  -0.1069 -0.2530  0.0870  0.2432  0.2602 -0.5945 -0.0845  0.2340 -0.4331
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7246 -0.6092  0.6597  0.4143 -0.7027  0.4462 -0.0101 -0.1381 -0.1556
#>   0.4881 -0.4388 -0.2692  0.6159 -0.7698  0.2603 -0.0115 -0.8169 -0.6299
#>   0.7319  0.7173  0.7898 -0.4749 -0.0619 -0.4307  0.6142  0.0643  0.1268
#> 
#> Columns 10 to 18 -0.1171 -0.2738  0.3001 -0.6376 -0.2373  0.3667  0.2291 -0.0866  0.2042
#>  -0.3726  0.2384  0.3201  0.1390  0.2783 -0.2724 -0.4756 -0.3401  0.4350
#>  -0.1237 -0.1210  0.1029  0.4649 -0.8675 -0.4416 -0.3544 -0.3818  0.4908
#> 
#> Columns 19 to 20  0.5320 -0.5621
#>   0.0109  0.4156
#>   0.3963 -0.0413
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0722 -0.0339  0.3132 -0.1543  0.3781 -0.0006  0.1087  0.3319 -0.0478
#>  -0.1054  0.3763  0.3470  0.5498  0.6555  0.0216  0.4712  0.0711 -0.1185
#>  -0.0801 -0.1309 -0.0724 -0.0265  0.0861 -0.3076 -0.0867  0.1865 -0.3019
#> 
#> Columns 10 to 18  0.5156 -0.0944  0.0314 -0.3316 -0.0176  0.4059 -0.1539 -0.1587 -0.0485
#>   0.2080  0.0305  0.3443 -0.5721  0.2064  0.0359 -0.5146 -0.4268  0.2671
#>   0.4834 -0.2184  0.2212 -0.2785  0.0138  0.6369 -0.0256 -0.1829  0.2311
#> 
#> Columns 19 to 20 -0.1531  0.5803
#>  -0.0955  0.5859
#>   0.1049  0.2264
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>