RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5989  0.2957 -0.1243  0.2297  0.4190  0.0923 -0.3774 -0.0215 -0.7021
#>  -0.4730  0.6302  0.3306 -0.4940 -0.8373  0.7590  0.0214  0.0241 -0.6852
#>  -0.1395  0.0086 -0.4450 -0.7514 -0.3980 -0.4659  0.1816  0.2920  0.6285
#> 
#> Columns 10 to 18  0.5654 -0.3137  0.5514 -0.7215  0.0412  0.2336  0.3780  0.4824 -0.6589
#>  -0.6072 -0.7580  0.2274  0.0964 -0.4010  0.4933 -0.3592  0.0195  0.0702
#>   0.2139  0.4007  0.5400 -0.5920 -0.7884  0.5619 -0.5263 -0.0805  0.5610
#> 
#> Columns 19 to 20  0.7409 -0.7547
#>   0.3340  0.1086
#>   0.2469 -0.8496
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0343  0.0229  0.1055 -0.5453  0.1070 -0.5691  0.4358 -0.1567 -0.0779
#>   0.2254  0.2567 -0.1089 -0.0187 -0.0350  0.4639  0.4906  0.3017  0.2213
#>  -0.0104  0.6335 -0.3089  0.1280 -0.4573  0.3442  0.3472 -0.4708 -0.1030
#> 
#> Columns 10 to 18 -0.0023 -0.3629  0.2641 -0.2930 -0.8998 -0.0314 -0.5300  0.4358  0.5464
#>   0.1735  0.4461 -0.1873  0.3147 -0.5579 -0.2789  0.2588  0.5765  0.2755
#>   0.3560  0.4386  0.5991  0.1765 -0.3031 -0.1037 -0.6713  0.5052  0.6824
#> 
#> Columns 19 to 20  0.6553 -0.6297
#>   0.4541  0.1198
#>   0.3776  0.0493
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1264 -0.2446 -0.3767  0.1966 -0.1583 -0.2978  0.1943 -0.0722 -0.1964
#>   0.2984  0.2374 -0.3159  0.0114 -0.4395  0.2430  0.4331  0.1026 -0.1785
#>   0.0311 -0.1574 -0.1128 -0.0093 -0.4735  0.2893 -0.0626 -0.2297  0.0133
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0587 -0.7343  0.0124  0.0059 -0.4683 -0.5624  0.2175  0.1037  0.4276
#>  -0.6446 -0.2709 -0.7113 -0.1495  0.1284 -0.5048  0.4884 -0.3941  0.3460
#>   0.2937 -0.3190  0.1053 -0.1691  0.4354 -0.3800  0.0611 -0.0893 -0.2615
#> 
#> Columns 10 to 18 -0.4173 -0.3309 -0.7235  0.1182  0.0943 -0.4744  0.7930 -0.7133 -0.0587
#>   0.1476  0.2329  0.1946  0.2874 -0.5828 -0.4748  0.7487 -0.2341  0.4197
#>  -0.4980 -0.2040 -0.4267 -0.2940  0.7418  0.5306 -0.3063 -0.2971 -0.0485
#> 
#> Columns 19 to 20 -0.4346  0.3689
#>   0.0562  0.1790
#>  -0.3836  0.5748
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0473 -0.1984 -0.5124 -0.0301 -0.1676  0.2368  0.4706  0.1464 -0.2120
#>  -0.3813  0.0100 -0.2307 -0.4054 -0.4220  0.2526  0.3261  0.0899 -0.2026
#>   0.0934 -0.1260 -0.2039  0.1224  0.1485 -0.2281  0.4545 -0.1447 -0.2133
#> 
#> Columns 10 to 18  0.7249  0.3136  0.0677  0.0100 -0.2283  0.0910 -0.2958  0.4244  0.6036
#>   0.5029  0.5956 -0.0148  0.0641  0.0953  0.4063 -0.5869  0.3121  0.3933
#>   0.6377 -0.2413  0.4793 -0.4172 -0.1983  0.0624 -0.5969  0.4772  0.3858
#> 
#> Columns 19 to 20  0.3105 -0.3935
#>   0.6592 -0.3018
#>   0.3774 -0.6695
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>