RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8444  0.3614  0.6294 -0.7994 -0.1800 -0.5100  0.0843 -0.4269  0.3348
#>  -0.3851 -0.1892  0.4673  0.7187 -0.3402 -0.2745 -0.2859 -0.6752 -0.4074
#>  -0.2866  0.3777 -0.0717 -0.6013  0.5173  0.1325 -0.2088 -0.3725  0.0147
#> 
#> Columns 10 to 18  0.3750 -0.3085 -0.4175 -0.2639  0.6507 -0.0316 -0.1766 -0.8093 -0.3621
#>  -0.4769  0.5351  0.3989 -0.7314 -0.4862  0.5118 -0.4489  0.5653  0.6156
#>   0.4512  0.3694 -0.7950 -0.5462 -0.7102 -0.1350  0.2867 -0.4565  0.4781
#> 
#> Columns 19 to 20 -0.8685 -0.4209
#>   0.4257  0.3665
#>   0.0315  0.3490
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1803 -0.0722 -0.1807  0.0267  0.4943  0.0743 -0.0189 -0.0099 -0.2142
#>   0.6556 -0.3342 -0.2086 -0.1251 -0.6563  0.1532 -0.1225  0.2551 -0.0932
#>   0.0405  0.0140  0.0566  0.2096 -0.0732  0.1146 -0.5104  0.6451 -0.0225
#> 
#> Columns 10 to 18  0.0962 -0.0734 -0.4159  0.3581  0.0700 -0.6904  0.2161  0.0174  0.6786
#>  -0.3930 -0.2854 -0.1456 -0.1559 -0.1612  0.6465 -0.2713 -0.3927  0.0062
#>   0.2242 -0.1377 -0.1437  0.0168  0.2761 -0.1431 -0.4011 -0.3815  0.3176
#> 
#> Columns 19 to 20 -0.4901 -0.2610
#>  -0.7246  0.3258
#>  -0.3808 -0.0242
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2103  0.2419  0.3840  0.0941 -0.0178  0.3493  0.1083  0.4948 -0.2088
#>   0.2560  0.1547  0.2801 -0.0220 -0.1552 -0.0429 -0.5034  0.3973 -0.2658
#>   0.1808  0.0144 -0.0614  0.1920 -0.0976 -0.2125 -0.3727  0.0801 -0.1447
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1929 -0.1356 -0.7687 -0.2357  0.0719 -0.6358 -0.7679  0.7231  0.0838
#>   0.2610 -0.5541 -0.6480 -0.0347 -0.8048 -0.4838 -0.3337 -0.1115  0.2518
#>  -0.4816 -0.4345 -0.5158 -0.2943 -0.1043 -0.0308  0.1278 -0.8128 -0.4668
#> 
#> Columns 10 to 18 -0.2989 -0.4269 -0.0209 -0.4430  0.4748 -0.1507  0.4004 -0.6967 -0.6479
#>   0.2496  0.8217 -0.6574 -0.5153  0.6202  0.8766 -0.0524  0.4106 -0.3923
#>  -0.0418 -0.1196  0.3715 -0.1118 -0.0857 -0.5556 -0.0802 -0.2144 -0.3062
#> 
#> Columns 19 to 20  0.5743 -0.6616
#>   0.0717 -0.0175
#>   0.5166  0.0254
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2838 -0.1130  0.6757 -0.0583  0.0653 -0.1304 -0.4147  0.4679 -0.2266
#>   0.1590  0.1543  0.1358  0.1259 -0.6627 -0.4369 -0.5189  0.4247 -0.2791
#>  -0.1178 -0.0008 -0.2403 -0.0857  0.0325 -0.0690 -0.3260 -0.0713  0.0869
#> 
#> Columns 10 to 18  0.2154 -0.4292  0.4080  0.3645  0.3346  0.0093 -0.3797 -0.0431  0.2818
#>   0.3112 -0.1074  0.0907  0.2766  0.5679 -0.0984 -0.5901 -0.2185 -0.0890
#>  -0.0800 -0.1961  0.1939  0.2213  0.2944 -0.1552 -0.3601 -0.3411 -0.0954
#> 
#> Columns 19 to 20 -0.6037 -0.0713
#>  -0.4188  0.3295
#>  -0.3565  0.0165
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>