RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9667  0.8354 -0.5502 -0.7840 -0.0880 -0.7180  0.1189  0.8414 -0.9118
#>  -0.2285  0.7725  0.0392 -0.2309 -0.8023 -0.7431  0.0490  0.3860 -0.1712
#>  -0.1734  0.9216 -0.7589 -0.5051  0.1214 -0.1188  0.3819  0.5807 -0.0768
#> 
#> Columns 10 to 18  0.4444 -0.5854  0.6271  0.3516  0.9380  0.8824 -0.6900  0.0008 -0.3769
#>  -0.4635 -0.1732  0.0746  0.5107  0.4933 -0.5409 -0.5433  0.5216 -0.5276
#>   0.7101 -0.8316  0.2233 -0.6976 -0.0532  0.8451 -0.2477  0.7540 -0.3504
#> 
#> Columns 19 to 20 -0.0420  0.3741
#>   0.3730 -0.3079
#>  -0.5230 -0.4441
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2040  0.8342 -0.2785 -0.2865 -0.1531  0.8535  0.5881  0.4881 -0.0001
#>  -0.5483  0.5224  0.4319 -0.2101 -0.0473  0.2684  0.2273  0.3272 -0.3934
#>  -0.2744  0.1773 -0.3293  0.1815 -0.2192  0.0830  0.1691  0.4949  0.0590
#> 
#> Columns 10 to 18 -0.3507 -0.2940 -0.2443 -0.6083  0.1437  0.9087  0.0904 -0.2077 -0.5242
#>  -0.1279 -0.6256  0.1025 -0.1433  0.6816  0.4788 -0.0317  0.5544 -0.6385
#>  -0.1696 -0.1486 -0.0184  0.1274  0.2593  0.1108 -0.2053 -0.4895 -0.0385
#> 
#> Columns 19 to 20 -0.1203 -0.3178
#>  -0.4245 -0.4892
#>   0.1165 -0.2204
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1675  0.3066  0.0770 -0.3868  0.1731  0.3267  0.1769  0.4205 -0.2909
#>  -0.0179  0.2123 -0.2006  0.4629 -0.1628  0.4823  0.2536  0.2633 -0.0911
#>  -0.0148  0.4539  0.3942 -0.0267 -0.4713 -0.0882  0.0160  0.7088 -0.1885
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0777  0.2344  0.2068  0.6205 -0.0693  0.2288 -0.5559  0.8454  0.1347
#>   0.4082  0.4282 -0.4794  0.4527  0.3017 -0.6289 -0.4595  0.4905  0.4121
#>  -0.4866  0.7037  0.4239  0.3450  0.2601 -0.7411  0.6501  0.2012 -0.3230
#> 
#> Columns 10 to 18  0.5913 -0.2191 -0.2513 -0.7722  0.4004 -0.2266  0.4034  0.3749 -0.2408
#>   0.8194 -0.3457 -0.4228 -0.5356  0.6922 -0.0594 -0.0257  0.0311 -0.0208
#>  -0.5803  0.1203 -0.4037 -0.4565 -0.6181 -0.4117  0.7255  0.3132  0.1415
#> 
#> Columns 19 to 20  0.6744 -0.3255
#>   0.8448 -0.2713
#>   0.0657  0.6240
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2646  0.3911 -0.0182  0.3797  0.2348  0.4515  0.1190  0.5092 -0.3844
#>  -0.4658  0.3569 -0.3672 -0.0215 -0.0841  0.2587  0.5557 -0.1240 -0.2843
#>  -0.1759  0.3710  0.0566  0.4131 -0.2196  0.1528  0.2328  0.7731  0.1236
#> 
#> Columns 10 to 18  0.0133 -0.4410 -0.1643 -0.2939  0.4964  0.5599  0.0457  0.0682 -0.0339
#>  -0.0552 -0.1891 -0.2547  0.1580  0.5684  0.3233  0.3226  0.2996 -0.3013
#>   0.3282 -0.4173  0.0961  0.1825  0.4199  0.5442 -0.2069 -0.3801 -0.5580
#> 
#> Columns 19 to 20 -0.3325 -0.4165
#>  -0.4268 -0.1523
#>  -0.2291 -0.4956
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>