RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3886  0.7210 -0.3640  0.4266  0.3991 -0.4447 -0.1434 -0.7696  0.5700
#>   0.1949 -0.3347 -0.3229 -0.8969 -0.4640 -0.7231  0.1135 -0.5647 -0.0547
#>   0.6433 -0.0000 -0.2986 -0.0845  0.5594 -0.6386  0.8107 -0.1402 -0.7312
#> 
#> Columns 10 to 18 -0.0510  0.3295  0.3253 -0.5985  0.6165 -0.0263 -0.3078 -0.7178  0.1601
#>  -0.2378 -0.6054  0.8428  0.2009 -0.5203 -0.7564 -0.3698  0.6297  0.8255
#>   0.1634 -0.4358  0.5551  0.5225  0.2347 -0.0212 -0.0393 -0.5690 -0.0471
#> 
#> Columns 19 to 20  0.3759  0.4420
#>  -0.3091 -0.2583
#>  -0.2789 -0.7045
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2306  0.4707  0.3488  0.1004 -0.0084  0.7043  0.0690 -0.4994  0.1476
#>  -0.0651 -0.3478 -0.3253 -0.1582 -0.2678 -0.1111  0.6018 -0.6113  0.1079
#>  -0.2419 -0.4829 -0.4348 -0.3901  0.3302  0.1384  0.2034 -0.3668  0.4053
#> 
#> Columns 10 to 18  0.0945  0.1483  0.3309 -0.0809  0.1316 -0.4931 -0.1849 -0.6206 -0.0129
#>  -0.2012  0.2592  0.3641 -0.0621  0.3046 -0.1354 -0.3325 -0.3239  0.3203
#>   0.5646  0.7869  0.0368  0.0278  0.2778  0.0007 -0.2107 -0.3268 -0.2714
#> 
#> Columns 19 to 20 -0.3631 -0.0014
#>   0.0427  0.1533
#>   0.3665 -0.4682
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4652  0.1064  0.0319 -0.2402 -0.1577  0.2516  0.4011 -0.5271 -0.0939
#>  -0.0170 -0.1095 -0.3690 -0.3299  0.2068  0.3870  0.4529 -0.3255 -0.2148
#>  -0.0008 -0.0014 -0.0853 -0.6307  0.3552  0.6288  0.3744 -0.2055  0.1351
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0338  0.3192 -0.2252  0.0249 -0.2863 -0.2286 -0.6584 -0.6782  0.7285
#>   0.1306 -0.6564 -0.1708  0.2794  0.1606 -0.2189  0.2691 -0.4289  0.5146
#>   0.6223 -0.7580 -0.6398  0.1082 -0.5803 -0.3506  0.4904  0.1976  0.7173
#> 
#> Columns 10 to 18 -0.4092  0.1577 -0.7251  0.4198 -0.7001 -0.1132 -0.1463 -0.4305  0.4241
#>  -0.7763 -0.7295 -0.1338  0.1985 -0.4540 -0.2535 -0.5272  0.1903  0.4811
#>  -0.0137 -0.0825  0.0312  0.3656 -0.4281 -0.5310  0.0285 -0.3070  0.2890
#> 
#> Columns 19 to 20 -0.0787 -0.1474
#>   0.1325 -0.4687
#>  -0.2245  0.5424
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4199 -0.4735  0.0860 -0.6605  0.1959  0.2860  0.4144 -0.1975  0.3609
#>  -0.2928 -0.3922 -0.2166 -0.3373 -0.1228  0.4131  0.4042 -0.4624  0.2583
#>   0.1841  0.0258  0.1580 -0.3963  0.2959  0.5465  0.1100 -0.4861 -0.2581
#> 
#> Columns 10 to 18  0.5653  0.1617  0.0932  0.2661 -0.0271 -0.1524  0.3622 -0.4462 -0.1068
#>   0.2244 -0.0014  0.4680 -0.1248 -0.0118 -0.5866  0.1389 -0.6327 -0.0305
#>   0.0998  0.3586 -0.3176  0.0276 -0.4733 -0.3665  0.3533 -0.5041  0.1811
#> 
#> Columns 19 to 20  0.0200 -0.6341
#>  -0.0030 -0.3614
#>  -0.0097 -0.3679
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>