RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2556 -0.3492  0.0783  0.3209  0.0675  0.6402 -0.6011 -0.5199  0.6193
#>   0.3022 -0.4131 -0.7535  0.5508 -0.2423  0.2714 -0.1756 -0.4595  0.3505
#>  -0.2935 -0.3502  0.4922 -0.2830  0.4216 -0.5376  0.4087  0.8442 -0.1073
#> 
#> Columns 10 to 18  0.8917  0.3709  0.1017  0.3829  0.3677  0.0295 -0.0490 -0.0305  0.1008
#>   0.5531  0.2821  0.7042  0.1109  0.8586  0.3157  0.4710  0.5657 -0.6627
#>   0.4038  0.2571  0.0978  0.9304 -0.3939  0.6043 -0.0265 -0.6823  0.2396
#> 
#> Columns 19 to 20 -0.5835  0.7175
#>  -0.4315 -0.3703
#>  -0.4986  0.8743
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0785 -0.7213  0.0221  0.2322 -0.0377 -0.5095  0.4758  0.2931  0.1955
#>   0.3006 -0.5642 -0.4385  0.1360 -0.7349 -0.2130  0.5071  0.3561 -0.2215
#>   0.2900 -0.7600 -0.0592  0.1106 -0.1543 -0.4535 -0.6643 -0.3801 -0.1477
#> 
#> Columns 10 to 18  0.2890  0.4590  0.2385  0.3197  0.0389  0.4407 -0.1138  0.5127 -0.5938
#>  -0.4390  0.2515 -0.1001 -0.0889  0.0870  0.5147 -0.0424  0.3164 -0.5698
#>   0.7055  0.5038 -0.1047  0.4474  0.4165  0.0018  0.6946 -0.2577  0.3830
#> 
#> Columns 19 to 20 -0.4329  0.4840
#>  -0.0377  0.1869
#>  -0.0382  0.7633
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4389 -0.6176 -0.5208  0.2633 -0.5214 -0.3632  0.1317  0.1145 -0.3389
#>   0.5891 -0.5253 -0.6143  0.2433 -0.4702 -0.5158  0.0218  0.1187  0.1480
#>   0.2617 -0.1691 -0.2570 -0.2065 -0.4476  0.2241 -0.4094  0.2387 -0.3481
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7508 -0.2920  0.1265  0.0382 -0.3844 -0.5680 -0.3526 -0.6145 -0.5648
#>  -0.2114 -0.1107  0.3051 -0.3175 -0.6816 -0.4703  0.6350  0.4814 -0.6389
#>   0.1042 -0.2481  0.2332  0.6497 -0.3884 -0.2110  0.3597  0.1290 -0.6543
#> 
#> Columns 10 to 18  0.1161 -0.4456  0.0087  0.1855  0.5723 -0.5602 -0.4472 -0.1893 -0.4266
#>  -0.0322 -0.3195  0.0241  0.4599  0.7757  0.2086  0.0112  0.4993 -0.5091
#>  -0.5319  0.2331  0.4453  0.1158  0.1836  0.2023  0.3982  0.2327 -0.1432
#> 
#> Columns 19 to 20  0.3071 -0.5130
#>  -0.0850  0.4009
#>  -0.4298 -0.4056
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5101 -0.4853 -0.6480  0.0608 -0.3415 -0.4387 -0.0174  0.1875  0.0946
#>   0.0452 -0.6613 -0.2143  0.1661 -0.3335 -0.1094  0.0750 -0.1541  0.3108
#>   0.1889 -0.6557 -0.6749  0.1859 -0.1792  0.0438  0.1203 -0.1457 -0.0643
#> 
#> Columns 10 to 18  0.4846  0.3630 -0.2557 -0.1654  0.3044  0.1943  0.1669  0.3199 -0.3437
#>   0.4365  0.2944  0.0234  0.1348  0.3676  0.1382  0.0882  0.4065 -0.6393
#>   0.0550  0.0684 -0.0729  0.0828  0.5597  0.4250 -0.0019  0.3739 -0.4226
#> 
#> Columns 19 to 20 -0.3920  0.7953
#>  -0.2259  0.7206
#>  -0.1951  0.3864
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>