RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8547  0.4764  0.5979  0.2511 -0.3870  0.3473 -0.3724 -0.3746  0.2885
#>  -0.3615 -0.6304  0.9223  0.4567 -0.6763 -0.8123  0.3156  0.3222  0.1784
#>   0.4063 -0.3849  0.6170  0.5091  0.6905 -0.8290  0.9445  0.8713 -0.3881
#> 
#> Columns 10 to 18  0.0948  0.1640 -0.1217 -0.8809 -0.7185 -0.1461  0.2942 -0.2150 -0.6430
#>   0.1146  0.0766 -0.3089 -0.7613 -0.7200  0.2841  0.3804  0.7547  0.7131
#>   0.6478 -0.7406 -0.4688  0.8812  0.1567  0.7623 -0.1617  0.7563  0.6283
#> 
#> Columns 19 to 20 -0.6951 -0.2988
#>   0.5043  0.3202
#>   0.4881 -0.4681
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1360 -0.3279 -0.1848  0.4578 -0.3933 -0.1045 -0.1258  0.3106  0.1040
#>  -0.1972 -0.2637 -0.1991  0.0596 -0.6217 -0.1698  0.3137 -0.0448  0.2206
#>  -0.6297  0.5089  0.2823  0.3234 -0.3967 -0.0557  0.1638 -0.1373 -0.1711
#> 
#> Columns 10 to 18 -0.0839  0.3604  0.3082 -0.1836  0.1243  0.7169  0.2214 -0.0700  0.0522
#>  -0.4079  0.1677  0.0191  0.1116  0.4846  0.0856 -0.3243  0.5793  0.0120
#>  -0.3069 -0.3568 -0.1991 -0.4380  0.1445 -0.7930 -0.6038  0.4829 -0.0652
#> 
#> Columns 19 to 20  0.4337  0.5316
#>   0.0875  0.4066
#>  -0.2615  0.0653
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3591  0.1441 -0.2919 -0.2173 -0.6204  0.5254  0.0088 -0.2060  0.4165
#>  -0.1054 -0.4555 -0.2058  0.5360 -0.1939 -0.1144 -0.0405 -0.5153  0.4370
#>  -0.4391 -0.4753 -0.0399  0.5512 -0.3769 -0.0921  0.0202  0.0009  0.3279
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8466 -0.1414  0.1139  0.3552  0.3115 -0.1197 -0.0232 -0.6434 -0.2701
#>  -0.2553  0.0140 -0.4928 -0.4554  0.7398 -0.4255  0.1791  0.3334 -0.3071
#>  -0.6297 -0.6498  0.0949 -0.2462 -0.0427  0.4230 -0.3023 -0.3357  0.6503
#> 
#> Columns 10 to 18  0.7006  0.4531  0.7116 -0.4605 -0.4222 -0.3277 -0.4717 -0.3096 -0.2244
#>  -0.3321  0.0731  0.1742 -0.1567 -0.5202  0.1096  0.0388  0.1738 -0.6785
#>  -0.4954  0.0248  0.4919  0.4302 -0.0784 -0.4240  0.0798  0.3158  0.0577
#> 
#> Columns 19 to 20  0.6192  0.4012
#>  -0.6336  0.3604
#>   0.1190 -0.7892
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2498  0.0642 -0.1186 -0.2078 -0.0837  0.0140  0.2026 -0.1473 -0.0007
#>  -0.3323  0.0956  0.0516  0.1201 -0.6073  0.2146 -0.2453 -0.4507  0.3998
#>  -0.3427 -0.0388 -0.6019  0.1654 -0.0575  0.5270  0.0221 -0.2140  0.6033
#> 
#> Columns 10 to 18  0.0949 -0.2370 -0.1335  0.3203  0.0982  0.5024  0.0482  0.3793  0.3608
#>  -0.2730 -0.2715 -0.1168 -0.3491  0.0862 -0.2391 -0.2604  0.1017  0.1803
#>   0.3625 -0.4572  0.4574 -0.2419  0.2822  0.1681  0.0667  0.2733 -0.0088
#> 
#> Columns 19 to 20 -0.0100  0.2397
#>   0.1356  0.2734
#>  -0.3711 -0.3601
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>