RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2791  0.5965  0.5361 -0.3908  0.8374 -0.1612  0.5724  0.4050  0.6011
#>   0.4868 -0.9357  0.8311  0.6711  0.5988  0.8437 -0.1995 -0.8523  0.1930
#>  -0.2634 -0.2710 -0.3144  0.5732  0.9406 -0.2346 -0.1346 -0.5138  0.9181
#> 
#> Columns 10 to 18 -0.0891 -0.8654  0.4238  0.0038  0.1878  0.1908 -0.8487 -0.4337  0.1827
#>  -0.8928 -0.0806  0.2063  0.5568  0.6771  0.7371 -0.6025  0.5811 -0.4368
#>  -0.2344 -0.8420  0.8359  0.7572  0.8592 -0.1899 -0.9133  0.5532 -0.2947
#> 
#> Columns 19 to 20  0.3217  0.6924
#>   0.2660  0.3330
#>   0.6972  0.2888
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3625 -0.4520  0.3559 -0.5043 -0.1204  0.4997 -0.1615  0.2502  0.3962
#>  -0.4659  0.3467  0.1748 -0.3054  0.7332 -0.2453 -0.4831 -0.0163  0.5214
#>   0.4983 -0.0299  0.2012  0.0740  0.2829  0.3889 -0.3493 -0.0516  0.5860
#> 
#> Columns 10 to 18 -0.2947 -0.7816  0.1928  0.4812 -0.6478  0.4840 -0.0423 -0.0738  0.2987
#>  -0.5468 -0.5034 -0.1449  0.0740  0.1102 -0.3215 -0.8277  0.0230 -0.0420
#>  -0.5279 -0.7440 -0.7575 -0.0779 -0.3380 -0.4465 -0.1876 -0.3268 -0.0474
#> 
#> Columns 19 to 20 -0.0020 -0.1011
#>   0.5725 -0.2862
#>   0.1794 -0.0984
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5661 -0.0008  0.5401 -0.6234  0.1614  0.6476 -0.3699 -0.2586  0.3045
#>   0.4199 -0.6877  0.5465  0.0249 -0.7222  0.6313 -0.7301 -0.4632  0.2882
#>  -0.2468 -0.4616  0.2873  0.3947  0.3164  0.4919 -0.2598  0.0722  0.4173
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1441  0.3800 -0.4399  0.0953 -0.7728  0.3577 -0.8206  0.0485 -0.6267
#>   0.0039  0.0584 -0.3534 -0.0304 -0.1205  0.0474 -0.4776 -0.0424 -0.4612
#>   0.3217 -0.6747 -0.6263 -0.8492  0.3539  0.0213 -0.3208 -0.4070  0.2235
#> 
#> Columns 10 to 18  0.5847  0.2444  0.4942  0.7048  0.1642 -0.4696 -0.0304 -0.6515  0.3677
#>  -0.0124 -0.2141  0.0628  0.1215  0.4734 -0.1429 -0.0522  0.5625  0.3196
#>  -0.2146  0.7249 -0.1274 -0.1305  0.2952 -0.1408 -0.2395  0.3113 -0.0532
#> 
#> Columns 19 to 20 -0.2421  0.6341
#>  -0.0651  0.6967
#>   0.8962 -0.2923
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3762  0.0082  0.6090 -0.1084  0.2537  0.3063 -0.1072 -0.0084  0.5386
#>  -0.2981  0.0229  0.3335 -0.0867  0.6026  0.1686 -0.1402  0.0181  0.3950
#>   0.4323 -0.2089  0.7933 -0.0392  0.2750  0.1768 -0.6800 -0.6576 -0.1967
#> 
#> Columns 10 to 18 -0.4507 -0.4746  0.1745  0.4504 -0.4093  0.4965 -0.1565  0.4049  0.3842
#>  -0.3099 -0.3173 -0.3284  0.0026 -0.3618 -0.3532 -0.4950 -0.3317 -0.1103
#>  -0.3241  0.2235 -0.3549 -0.1246  0.3226 -0.2855  0.0227 -0.2614 -0.3855
#> 
#> Columns 19 to 20  0.1989 -0.2283
#>   0.2667 -0.1665
#>   0.3941  0.1715
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>