RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5481  0.9255  0.4442 -0.1352  0.4433 -0.2590  0.0491  0.3924  0.0395
#>   0.0015  0.0840 -0.7086 -0.4587  0.1391  0.5713  0.4221  0.0029 -0.3624
#>  -0.6475  0.8430 -0.7086 -0.5409 -0.2828  0.7860  0.0390  0.2899 -0.4603
#> 
#> Columns 10 to 18 -0.4494 -0.0754  0.7496  0.0069 -0.2542  0.1027 -0.1921 -0.0431 -0.6809
#>  -0.5166 -0.0013  0.1876 -0.5080  0.2332  0.7803 -0.5562 -0.6688 -0.6761
#>  -0.5933 -0.1197  0.3035  0.6344  0.0596  0.0713  0.2173  0.2542  0.5214
#> 
#> Columns 19 to 20 -0.8027 -0.6911
#>  -0.5698  0.0735
#>   0.4408 -0.6127
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0568  0.7169  0.0052 -0.6275 -0.0277 -0.5777  0.1861  0.8079 -0.0204
#>   0.2136  0.6434 -0.2629 -0.1033 -0.2267  0.3005 -0.0410  0.3581 -0.1920
#>   0.3204  0.2847  0.1245 -0.1290 -0.6154  0.0764 -0.2516 -0.1289 -0.3169
#> 
#> Columns 10 to 18 -0.1946 -0.0754  0.5055  0.4278  0.3852  0.5650 -0.8219 -0.1246 -0.3161
#>  -0.7276  0.1622 -0.2643  0.0891  0.0971  0.4766 -0.4940 -0.5019  0.0072
#>  -0.6071  0.0345 -0.1984  0.2886  0.5243  0.5234 -0.4842 -0.2201 -0.6016
#> 
#> Columns 19 to 20 -0.8130 -0.3246
#>  -0.3499  0.1533
#>  -0.5084  0.0009
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2150  0.8247 -0.2068 -0.4442 -0.1727 -0.0007  0.0798  0.6533 -0.2106
#>   0.2603  0.6060 -0.0549 -0.2483 -0.2752  0.3036  0.1185  0.4063 -0.1355
#>   0.1145  0.7429 -0.6588 -0.3393  0.3035  0.1070 -0.1059  0.0168  0.0786
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3323  0.3254  0.1297 -0.4414 -0.4141  0.3812  0.0203  0.2430  0.4708
#>  -0.8605  0.0363  0.2545  0.1363 -0.4618  0.5935  0.3382 -0.1289 -0.8537
#>  -0.2210  0.5121  0.0448 -0.0824 -0.3595 -0.1422  0.4925  0.7115 -0.1725
#> 
#> Columns 10 to 18 -0.4714 -0.8039 -0.8408 -0.3761 -0.0110  0.5964  0.0221 -0.2341  0.5049
#>  -0.5244 -0.1433 -0.4727 -0.4543  0.5905  0.6291  0.7385 -0.2575  0.5907
#>  -0.3747  0.0466 -0.0935 -0.1365  0.5030  0.4207 -0.2433 -0.3627  0.3983
#> 
#> Columns 19 to 20  0.0074 -0.3910
#>  -0.1992 -0.1914
#>  -0.1488  0.3164
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2957  0.6238 -0.4617 -0.2342 -0.5110 -0.2387 -0.0794  0.1232  0.3349
#>   0.2719  0.5641 -0.2478 -0.2679 -0.2292 -0.0333 -0.3576  0.0803  0.1154
#>   0.0757  0.7010 -0.2375 -0.2406  0.0004 -0.0584  0.1520  0.3291 -0.0002
#> 
#> Columns 10 to 18 -0.5264 -0.1507  0.0766  0.2247  0.2213  0.1329 -0.6741 -0.1747 -0.4136
#>  -0.2335 -0.2819  0.3287  0.4416  0.6214  0.0855 -0.5895  0.4680 -0.4400
#>  -0.0235 -0.2800  0.4159  0.4859  0.4008  0.0412 -0.4453  0.2288 -0.0393
#> 
#> Columns 19 to 20 -0.4668  0.1834
#>  -0.2064  0.0517
#>  -0.0753 -0.0710
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>