RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2206  0.8265 -0.5981 -0.1365 -0.5813 -0.6520  0.0762 -0.6895  0.4307
#>  -0.4554  0.3422 -0.3970 -0.2347 -0.5970 -0.9257 -0.2734 -0.2927  0.0110
#>   0.1642 -0.3550 -0.9051 -0.8951  0.7473 -0.4107  0.5473 -0.1392  0.8045
#> 
#> Columns 10 to 18 -0.2420 -0.9266  0.2968 -0.0776 -0.9281 -0.3953  0.2809  0.8794 -0.0441
#>  -0.2650 -0.7212 -0.2501 -0.5356 -0.2244  0.6055  0.1282  0.3130  0.6120
#>   0.7862 -0.9058  0.2498  0.9361 -0.8446 -0.5596 -0.6491 -0.2200 -0.2423
#> 
#> Columns 19 to 20 -0.7782  0.2812
#>  -0.4454  0.1967
#>  -0.5791 -0.4285
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1751  0.0092  0.3867 -0.4225 -0.2261  0.1958 -0.4420 -0.1113  0.3465
#>  -0.0917 -0.2429  0.6201 -0.2285 -0.7046  0.2555 -0.2524  0.3402  0.6225
#>  -0.1619  0.7814 -0.1219 -0.7494 -0.1567 -0.0301 -0.0879  0.0145  0.4835
#> 
#> Columns 10 to 18 -0.2823 -0.1465  0.1583 -0.0599 -0.1207 -0.4484  0.1237 -0.0929 -0.3693
#>   0.4268  0.1237  0.1718 -0.5089  0.0528 -0.3317 -0.2705 -0.3484 -0.0963
#>  -0.5321  0.3438  0.6458  0.0308 -0.7873 -0.3227  0.1733 -0.0987 -0.4211
#> 
#> Columns 19 to 20  0.0765  0.6678
#>   0.5363  0.6064
#>   0.2852  0.5656
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1473  0.4061 -0.2078 -0.6196 -0.7317 -0.1512 -0.6334 -0.0292  0.5335
#>   0.2798  0.5068 -0.5515 -0.4036 -0.5501 -0.2662 -0.4918 -0.2146  0.6602
#>  -0.0780  0.3821 -0.1309 -0.1755  0.0774 -0.1962 -0.4953 -0.2508 -0.1650
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2811  0.4740  0.3112 -0.6013  0.1470 -0.0360  0.2243  0.4757 -0.3597
#>   0.4723 -0.6364  0.1489  0.0728  0.0616 -0.0962 -0.6390 -0.3158 -0.3413
#>   0.5582  0.2833 -0.0413 -0.1857  0.4345  0.1982  0.1263  0.2894  0.4490
#> 
#> Columns 10 to 18 -0.5060  0.2292  0.3020  0.2893 -0.0149 -0.8178 -0.5008 -0.3480  0.6524
#>  -0.5127 -0.1080  0.3427  0.0769  0.2472 -0.4644  0.2523  0.2976 -0.0280
#>   0.7627  0.3182 -0.6528  0.5012 -0.0323 -0.8601 -0.7799 -0.5046  0.6776
#> 
#> Columns 19 to 20 -0.3727 -0.0021
#>  -0.3825 -0.1883
#>   0.6292 -0.5001
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3071  0.0137  0.1624 -0.4071 -0.6669 -0.2631 -0.3565 -0.1948  0.2856
#>   0.0113 -0.1236 -0.2292 -0.3331 -0.7279 -0.2413 -0.3383 -0.2932  0.3683
#>  -0.3915  0.3907  0.0104 -0.7283 -0.5649  0.1260 -0.5499  0.2425  0.0052
#> 
#> Columns 10 to 18 -0.4134 -0.3492  0.0170  0.2522 -0.3506  0.0187  0.4790  0.1135  0.0563
#>  -0.0975  0.1338 -0.1367 -0.0369 -0.0625 -0.2063  0.2188 -0.1199 -0.2079
#>  -0.3049 -0.6756  0.1603 -0.0507 -0.6369  0.0095  0.0561  0.1067 -0.0448
#> 
#> Columns 19 to 20  0.1086  0.6300
#>   0.2022  0.0583
#>   0.4662  0.8990
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>