RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3061  0.0315  0.6697  0.4435 -0.3318  0.0762  0.8668  0.1960 -0.3247
#>  -0.5316  0.2018  0.3006  0.6808  0.8251 -0.0294 -0.9009 -0.6115  0.7326
#>  -0.3500  0.2152  0.5738 -0.5194 -0.1679 -0.4333 -0.3744  0.0363 -0.8739
#> 
#> Columns 10 to 18 -0.1109 -0.1510  0.1516  0.3852  0.4166 -0.0471 -0.3341 -0.3451  0.6802
#>  -0.1445  0.6105  0.2658 -0.0687  0.7095  0.6887 -0.1093 -0.6288 -0.4974
#>  -0.6904 -0.0709  0.3227  0.2878  0.0271 -0.2137 -0.6450 -0.0556 -0.9067
#> 
#> Columns 19 to 20 -0.1027  0.6562
#>   0.5538 -0.0635
#>   0.0276  0.0553
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1075 -0.4908  0.5727 -0.1763  0.1260 -0.1558 -0.6996  0.6198  0.0209
#>   0.2744  0.0168  0.6104  0.2565  0.3638  0.0631 -0.5881  0.1666 -0.2034
#>  -0.2576 -0.0402  0.4585 -0.3650  0.2707 -0.1903 -0.1486  0.4905 -0.7321
#> 
#> Columns 10 to 18 -0.6644  0.1170  0.3347 -0.0165  0.1883 -0.6376 -0.4619  0.0262  0.0864
#>  -0.2073 -0.3211 -0.6491  0.3139  0.5548 -0.1032 -0.3943  0.0291  0.1870
#>   0.1972  0.3554 -0.3666  0.6472  0.2061 -0.4209 -0.5599 -0.1110  0.2874
#> 
#> Columns 19 to 20  0.3066  0.4748
#>   0.0262  0.6135
#>   0.1237  0.3413
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1083 -0.0169  0.2307  0.1488  0.5138 -0.1820 -0.2561  0.1987 -0.1389
#>   0.0076 -0.1082  0.5277  0.0143  0.1454  0.0313 -0.1396  0.5423 -0.6811
#>   0.0741 -0.3396  0.0700 -0.0111  0.2941 -0.0708 -0.3028  0.6135 -0.1039
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0394  0.2181 -0.4192  0.6316  0.2322  0.3345 -0.5828  0.1486 -0.6036
#>   0.1000 -0.6359 -0.5373  0.1342  0.5628 -0.0350 -0.4335  0.0511  0.1155
#>  -0.3928  0.0683 -0.3148  0.1359  0.8905 -0.2818  0.2491  0.5804 -0.0901
#> 
#> Columns 10 to 18  0.3443 -0.5427  0.2965  0.4827  0.4706 -0.8890  0.1208 -0.1929 -0.2339
#>  -0.0699 -0.0966  0.1672 -0.0823  0.4682 -0.2113  0.5865 -0.2520 -0.1152
#>   0.2182  0.7171 -0.0217  0.7173 -0.1657 -0.5804 -0.0235 -0.4792  0.2971
#> 
#> Columns 19 to 20  0.1804 -0.0896
#>  -0.2678 -0.0110
#>   0.4813 -0.2666
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4580  0.0526  0.2248 -0.0203  0.3184 -0.1279 -0.3064  0.4959 -0.4304
#>  -0.3519 -0.0726 -0.1411  0.2520  0.3713 -0.2898 -0.2589  0.0026 -0.2252
#>  -0.2486  0.2647  0.1530  0.4296  0.3784  0.0873 -0.3464  0.5563 -0.1116
#> 
#> Columns 10 to 18  0.0562  0.3238 -0.3469  0.2899  0.1856 -0.0621  0.0148  0.3814  0.2315
#>   0.1723  0.2018 -0.5080  0.5018  0.2373 -0.2715  0.0330 -0.1155  0.2804
#>   0.1911 -0.1481 -0.1041  0.4083  0.2676 -0.0227 -0.0854  0.1566 -0.0695
#> 
#> Columns 19 to 20  0.2411  0.1281
#>   0.2090 -0.2593
#>   0.1199  0.6403
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>