RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7788  0.5183  0.0700  0.0346  0.4382  0.0305  0.2101  0.4402  0.4123
#>  -0.7506  0.9720 -0.8185 -0.8018  0.1160  0.9073  0.1004  0.4486 -0.1031
#>   0.6870  0.7975 -0.3385 -0.1724 -0.8870  0.8943  0.4771  0.5993 -0.4694
#> 
#> Columns 10 to 18 -0.6920 -0.3450  0.7548  0.4276  0.0939 -0.8306  0.2129  0.1482 -0.5468
#>  -0.3006  0.2227  0.5497 -0.4046  0.1485 -0.0773  0.3076 -0.2055  0.6924
#>  -0.7146 -0.4921 -0.3107  0.5309  0.2886 -0.7981 -0.1899 -0.2311 -0.4912
#> 
#> Columns 19 to 20 -0.3768 -0.1705
#>   0.3381  0.7730
#>   0.1665 -0.3955
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4785  0.3930 -0.6566 -0.6766 -0.2788  0.3183  0.5974 -0.1483 -0.0498
#>  -0.4209  0.3196 -0.4153  0.2813 -0.0239  0.2055  0.1205  0.4956  0.3871
#>  -0.1809  0.5513 -0.1384 -0.0362 -0.6857  0.2793  0.5184 -0.2583  0.2014
#> 
#> Columns 10 to 18  0.4110 -0.2157  0.0542 -0.5138  0.0857 -0.1548 -0.1068 -0.2458 -0.1486
#>   0.3411  0.3541 -0.0489  0.5257  0.5315 -0.4441 -0.3309  0.4368 -0.2061
#>   0.0192 -0.1406  0.2584  0.0216  0.6329 -0.0841  0.0758 -0.6846 -0.1563
#> 
#> Columns 19 to 20  0.2777 -0.1313
#>  -0.2738 -0.7044
#>   0.1873 -0.2964
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2025  0.1245 -0.4744 -0.2809 -0.2226 -0.3569 -0.1328  0.3948 -0.0772
#>  -0.3837  0.5759 -0.5871 -0.3207 -0.0039  0.4389  0.2392  0.6688 -0.5247
#>  -0.1143  0.7347 -0.1056 -0.0841 -0.2015  0.6057  0.0702  0.0756  0.0675
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4359  0.2321  0.2435 -0.2010 -0.3335  0.7329  0.0685  0.1729 -0.0781
#>   0.6103  0.1231  0.2151  0.0237 -0.1996  0.2783  0.1587 -0.6052 -0.4508
#>  -0.1058 -0.2298  0.6090  0.0506 -0.2323  0.3998 -0.2078  0.5096 -0.1042
#> 
#> Columns 10 to 18  0.3269 -0.4428  0.0688 -0.3315 -0.3976  0.1974 -0.3950 -0.6916 -0.4152
#>   0.5725 -0.2963 -0.3520  0.0929 -0.9072  0.4678 -0.8354 -0.6125 -0.4557
#>   0.5741  0.1157  0.5784  0.0442 -0.4903  0.1198  0.1359 -0.3439  0.2957
#> 
#> Columns 19 to 20 -0.6245 -0.2888
#>  -0.9115  0.0671
#>  -0.5068 -0.4619
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0116  0.3231 -0.5806 -0.6036 -0.3325  0.2561  0.4879  0.3246 -0.0912
#>  -0.1091  0.3907 -0.4603 -0.6999 -0.0433  0.0256  0.2413  0.3697 -0.2399
#>  -0.3532  0.3954 -0.5620 -0.3258 -0.3404  0.4045  0.4366  0.1602 -0.1809
#> 
#> Columns 10 to 18  0.4716  0.0384 -0.1433  0.0494  0.1021 -0.0377 -0.4165  0.3655  0.2153
#>   0.3953  0.1900 -0.0451  0.0723  0.4876 -0.3236 -0.1592  0.2467  0.5191
#>   0.3696  0.1718  0.1959 -0.1896  0.4778  0.2555 -0.4873 -0.2598  0.2270
#> 
#> Columns 19 to 20  0.4450 -0.2748
#>   0.1793 -0.3042
#>   0.1075 -0.1222
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>