RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3022  0.9923 -0.1493 -0.7630 -0.3890  0.0317  0.2457 -0.4111  0.9485
#>  -0.6499 -0.8634 -0.4287  0.8552  0.4152  0.0320  0.6800  0.8536 -0.2714
#>   0.1846 -0.2398  0.1531  0.4553 -0.8072 -0.0082 -0.4594 -0.5539 -0.1598
#> 
#> Columns 10 to 18 -0.7634  0.2073 -0.9130  0.5731  0.5953  0.1130  0.8970 -0.0714  0.1467
#>   0.0571  0.4754  0.4066  0.7620 -0.4829 -0.8763  0.0631  0.2441 -0.8007
#>   0.5814  0.7691 -0.0282 -0.4778  0.4575 -0.8328 -0.5546  0.5417  0.2132
#> 
#> Columns 19 to 20  0.4655 -0.6938
#>   0.8074  0.6327
#>  -0.8870 -0.4640
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6547  0.3083  0.2173  0.6288  0.7631  0.0902  0.7210  0.4229 -0.5070
#>   0.3059 -0.2534  0.2973  0.6817  0.0467 -0.2400  0.5440 -0.6692  0.0073
#>   0.6438  0.4197  0.3007 -0.2041 -0.1094  0.6241 -0.1011  0.1977 -0.5674
#> 
#> Columns 10 to 18  0.2024 -0.3194  0.4370  0.1975  0.6558 -0.5937 -0.7147 -0.2296  0.4529
#>   0.3994 -0.5120  0.3424  0.5407 -0.2502  0.1022 -0.2862 -0.2915  0.0224
#>   0.0589 -0.3589  0.3572 -0.1537  0.1944 -0.4748 -0.6588  0.3833 -0.2333
#> 
#> Columns 19 to 20  0.4242 -0.1180
#>  -0.1798 -0.0999
#>  -0.1058  0.3673
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.7515  0.1803  0.2441 -0.5419  0.6725 -0.1581 -0.2672  0.3132 -0.0374
#>   0.0745 -0.0652  0.0831 -0.3832  0.0587  0.0645  0.1906 -0.4193 -0.3634
#>   0.4790  0.2469  0.1169 -0.2668  0.1332 -0.0283 -0.1810 -0.0563 -0.5717
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3351 -0.1473 -0.8319 -0.1050  0.2289  0.8494 -0.3338 -0.0359 -0.5574
#>   0.1984  0.2456 -0.2479  0.1293  0.1831 -0.3724 -0.0030 -0.1714 -0.1119
#>  -0.3975 -0.3695 -0.5491  0.4444 -0.6913 -0.3930 -0.4314 -0.7896  0.4229
#> 
#> Columns 10 to 18 -0.5960 -0.3952  0.6153  0.2871 -0.2200 -0.9615  0.7221  0.5927 -0.7728
#>   0.3643  0.5138  0.2351  0.3235  0.0733 -0.3882  0.2376  0.4711 -0.6495
#>  -0.2902 -0.1917  0.0894 -0.8396 -0.5806  0.2661 -0.0032  0.2798  0.6281
#> 
#> Columns 19 to 20  0.0402 -0.0453
#>   0.6134 -0.1170
#>   0.3327  0.6210
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3482  0.1120 -0.1791 -0.2849  0.5543 -0.2685 -0.4427  0.5504  0.2638
#>  -0.1256  0.3630  0.0417 -0.2810  0.0170 -0.2621  0.0077 -0.3133 -0.1253
#>   0.1956  0.1527  0.1090 -0.1555  0.3933 -0.1057 -0.2678  0.4696 -0.0096
#> 
#> Columns 10 to 18  0.0041 -0.3818 -0.4389  0.0458 -0.0295  0.2432 -0.0821  0.0017 -0.0018
#>   0.3165 -0.1126 -0.2793  0.1433 -0.0458  0.2784 -0.2608 -0.0022  0.2856
#>   0.0949 -0.2589  0.1683  0.3717 -0.5006 -0.4587 -0.6615  0.3303  0.0108
#> 
#> Columns 19 to 20  0.3518 -0.0134
#>   0.0798 -0.3874
#>   0.7791  0.4536
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>