RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7833  0.4595  0.0539  0.2113  0.1698 -0.3869  0.8174  0.6560 -0.3211
#>   0.0429  0.1356  0.2820 -0.1848 -0.0880 -0.1450 -0.5191  0.1160  0.0131
#>   0.1512 -0.6968  0.8224 -0.6054 -0.2044  0.3279  0.6313  0.8062 -0.0217
#> 
#> Columns 10 to 18 -0.0105  0.4919  0.3090  0.2900  0.8188 -0.2877  0.5460  0.8024  0.1605
#>  -0.5156  0.1238  0.1877  0.4082  0.6254 -0.4046  0.5585  0.4917  0.1386
#>  -0.3297 -0.7552  0.6121  0.8561  0.7516 -0.8508  0.8419  0.2653 -0.6657
#> 
#> Columns 19 to 20 -0.8187  0.6747
#>   0.2138  0.5065
#>   0.0239 -0.5942
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3788  0.3539  0.1600 -0.2747  0.0836  0.1687  0.3403  0.1442  0.2239
#>   0.0410 -0.1554  0.1709  0.0213  0.0851  0.2421  0.1547  0.2741  0.5106
#>   0.4289 -0.4056  0.1900 -0.0729  0.0028  0.4229 -0.2923 -0.3588  0.7641
#> 
#> Columns 10 to 18  0.1665 -0.1355  0.1383  0.1799  0.0021  0.2991 -0.3497 -0.5812  0.1951
#>   0.3724 -0.3416 -0.3570  0.0734 -0.3972  0.3804  0.3609  0.1932  0.0237
#>  -0.0988 -0.0351 -0.6200  0.3586 -0.4345  0.0281  0.8301 -0.0686  0.3772
#> 
#> Columns 19 to 20 -0.6726  0.3628
#>   0.0251 -0.2685
#>   0.5678  0.2619
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1492 -0.2619  0.0492 -0.1529 -0.2561 -0.2352  0.3613  0.2553  0.2205
#>   0.3604 -0.4604  0.6872 -0.1202 -0.0381  0.2873  0.0274  0.0626 -0.0612
#>  -0.2390 -0.4665  0.6258 -0.5683 -0.0305  0.0175  0.5852  0.5235  0.4343
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1404  0.0381 -0.1720  0.0192 -0.0260 -0.3934 -0.5640 -0.2844 -0.7386
#>   0.1942 -0.5818  0.8257 -0.2316 -0.0121 -0.4451 -0.2537  0.4656 -0.5151
#>  -0.8156 -0.5731 -0.0793  0.2489  0.6261  0.0448 -0.2020 -0.3702  0.3305
#> 
#> Columns 10 to 18  0.5530 -0.6547  0.2088  0.0788 -0.6178 -0.1162  0.0663 -0.2448 -0.3806
#>   0.8251 -0.4248 -0.2283  0.3687 -0.1160 -0.2605 -0.1192  0.4116 -0.3948
#>   0.5631  0.4484 -0.4937  0.6616  0.2300 -0.5040  0.4723  0.0606 -0.8857
#> 
#> Columns 19 to 20 -0.1404 -0.6461
#>  -0.0466 -0.3419
#>  -0.3584  0.2013
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0323 -0.0143  0.4796  0.3032 -0.2672  0.1289 -0.2041  0.2738  0.4978
#>   0.5605  0.1251  0.5733  0.1635 -0.2619  0.0806 -0.3195  0.1103  0.1800
#>   0.5010  0.4677  0.4098 -0.5893 -0.6295 -0.1372  0.4300  0.2464 -0.1838
#> 
#> Columns 10 to 18 -0.1480  0.1916 -0.0663  0.3705  0.0204  0.2488  0.4977  0.0547 -0.2312
#>  -0.1702 -0.0879 -0.3051  0.4670 -0.1327  0.0228  0.4030 -0.2182  0.1948
#>  -0.6345 -0.2465  0.1575  0.2354  0.5154  0.0082  0.2606  0.2829 -0.4667
#> 
#> Columns 19 to 20 -0.1069  0.0853
#>   0.2450  0.2879
#>  -0.2896  0.1783
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>