RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4423 -0.5153 -0.8554  0.3764  0.4427 -0.0866  0.3216 -0.3440  0.0218
#>   0.3977 -0.8990 -0.3965  0.1883  0.1202  0.9517  0.1804  0.2629  0.6012
#>  -0.2122  0.0308  0.3868  0.4477  0.0575  0.3755  0.3706 -0.0360  0.8060
#> 
#> Columns 10 to 18  0.7056  0.6344  0.5533 -0.7055 -0.1589 -0.2623  0.0446  0.7503  0.5130
#>   0.8247 -0.7468  0.8371  0.7750 -0.2640 -0.2185 -0.5368  0.8498  0.0903
#>   0.4701 -0.2785  0.8586 -0.3891 -0.3365 -0.8597  0.1201 -0.7096  0.0605
#> 
#> Columns 19 to 20 -0.1584  0.8521
#>  -0.5240 -0.4670
#>  -0.6451 -0.2555
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2764 -0.5056  0.3997  0.0865 -0.4894  0.1998  0.6116  0.5924 -0.3112
#>   0.7589  0.2581 -0.0946  0.2104  0.2425  0.2219  0.5785  0.0909  0.4598
#>   0.7216 -0.2203 -0.0601 -0.3797 -0.4485 -0.1835  0.3919  0.0871 -0.1460
#> 
#> Columns 10 to 18  0.2484  0.3183  0.1684  0.3837  0.7302 -0.2372 -0.5059  0.3361 -0.2205
#>   0.3249 -0.3641 -0.1752  0.4063  0.3597  0.1778 -0.6692  0.2756 -0.1201
#>   0.1648  0.3127 -0.2484  0.1764  0.4003 -0.0959  0.4011  0.6138 -0.4369
#> 
#> Columns 19 to 20 -0.6982 -0.9613
#>   0.4119 -0.6250
#>  -0.1453 -0.1654
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4623 -0.1617 -0.2499  0.2681  0.1200  0.7635  0.2131 -0.3869  0.2649
#>   0.7084  0.3089  0.4269 -0.0748 -0.1710 -0.3399  0.2239 -0.5440  0.0412
#>   0.5901 -0.0633 -0.3459  0.4863  0.0565  0.6853  0.1959 -0.0985 -0.3937
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6957  0.2139  0.6247  0.1360  0.2832 -0.2024  0.5063  0.2120 -0.6012
#>  -0.6434 -0.5438 -0.6932  0.1551  0.6213 -0.3927 -0.4522 -0.1209 -0.6112
#>   0.0962  0.4975  0.8408  0.1748  0.1350 -0.4089  0.2452 -0.8375 -0.2020
#> 
#> Columns 10 to 18  0.3327  0.0434 -0.4926  0.0864  0.3432  0.2556 -0.0807  0.4000  0.3153
#>   0.2469  0.2100 -0.5327 -0.6528  0.3897  0.1377  0.2138 -0.1432  0.4127
#>  -0.3528  0.1097  0.6250  0.4757 -0.5692  0.0712 -0.8899 -0.4979  0.5277
#> 
#> Columns 19 to 20  0.3660 -0.1452
#>   0.7986 -0.5573
#>  -0.3830 -0.0198
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6703  0.0079  0.1041 -0.2091 -0.2140  0.0635 -0.0558 -0.1657 -0.4474
#>   0.5474 -0.3529 -0.1129 -0.0527  0.4080  0.2166  0.2730  0.0760  0.0536
#>   0.7866 -0.1354  0.1068 -0.0365 -0.3139 -0.2653  0.1216 -0.2306  0.2076
#> 
#> Columns 10 to 18  0.0177 -0.4522  0.4068  0.2270  0.3154 -0.1829 -0.1061  0.1223 -0.2600
#>   0.0362 -0.3858  0.2969  0.2583 -0.0198  0.1299 -0.1088  0.6486 -0.5292
#>  -0.2091 -0.2429 -0.3783  0.2498  0.5975 -0.1511  0.3536  0.2116 -0.0204
#> 
#> Columns 19 to 20  0.1322 -0.3949
#>   0.5842 -0.0934
#>   0.1687 -0.0576
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>