RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2114 -0.3182 -0.3384  0.7793  0.1229 -0.0988 -0.0783  0.9512  0.2319
#>   0.0638  0.8175 -0.6402 -0.9796  0.2137 -0.1274  0.2665 -0.2899 -0.1016
#>  -0.3587 -0.4396 -0.2543  0.0886  0.0732  0.5414  0.0273  0.5074 -0.1647
#> 
#> Columns 10 to 18 -0.3334 -0.2510  0.6275  0.2921 -0.9627 -0.9059 -0.2702 -0.4672 -0.3093
#>  -0.4211  0.0253  0.4512  0.4139 -0.1444 -0.4033  0.5459 -0.9418  0.8072
#>   0.6052 -0.2449  0.6973 -0.5579 -0.2940  0.1999 -0.7188  0.5921  0.0594
#> 
#> Columns 19 to 20 -0.9101 -0.3377
#>  -0.6056 -0.6102
#>   0.2558 -0.7055
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4047  0.2416 -0.2617 -0.0766  0.3529  0.2911 -0.2788  0.1606  0.2004
#>   0.1525  0.5079 -0.3698 -0.3952  0.0860 -0.3918 -0.2879  0.1040 -0.0242
#>  -0.2235  0.0752 -0.3313  0.1428  0.1609  0.1177  0.1771  0.1451  0.2242
#> 
#> Columns 10 to 18 -0.6327  0.5214  0.2192  0.2000 -0.6886 -0.4210  0.4082 -0.7974 -0.3228
#>  -0.1054  0.5282  0.0295  0.7543 -0.2236 -0.1450  0.4805 -0.3673  0.6132
#>  -0.6225 -0.4739  0.3550  0.3053 -0.3578 -0.6122  0.5063  0.6605 -0.5460
#> 
#> Columns 19 to 20 -0.3785 -0.3303
#>   0.4154  0.0701
#>   0.3998  0.1586
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3234  0.3951 -0.2656  0.0657  0.2941 -0.1866 -0.3379  0.2685 -0.2425
#>  -0.3551  0.2825  0.3879 -0.3769  0.2889 -0.3713 -0.2396  0.6799 -0.2573
#>  -0.1932 -0.1800  0.2361  0.0106  0.3243 -0.1274 -0.2056  0.1305  0.0687
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4001 -0.0692  0.1509  0.4891 -0.2015 -0.1158  0.4675 -0.0947 -0.1758
#>   0.7785  0.8004 -0.1667 -0.4453  0.6453  0.5567  0.0737  0.0721  0.2193
#>  -0.2792  0.6494 -0.1097  0.4735 -0.7094  0.7261  0.2986  0.4060  0.1443
#> 
#> Columns 10 to 18 -0.4959  0.1578 -0.2843 -0.2812  0.4058 -0.3783 -0.0367 -0.7899  0.3718
#>   0.7910  0.5110 -0.1451 -0.4314  0.9362 -0.4764 -0.0606 -0.4461 -0.0769
#>  -0.6549  0.2181 -0.4263 -0.0983  0.3236 -0.0851  0.2690 -0.4926  0.4953
#> 
#> Columns 19 to 20  0.2034  0.1292
#>  -0.5265 -0.7850
#>  -0.3588 -0.3050
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2811 -0.0830  0.2655 -0.0636  0.2482 -0.2540 -0.3593  0.2901 -0.0857
#>  -0.3519  0.1277  0.1425  0.1182  0.6796 -0.1616  0.0288  0.1743  0.1182
#>  -0.0477  0.1788 -0.0202 -0.4929  0.4136 -0.3793 -0.0764  0.1082 -0.3763
#> 
#> Columns 10 to 18 -0.5016  0.4265  0.0908  0.3742 -0.2180 -0.1757  0.3209 -0.2274 -0.0249
#>  -0.7245  0.5909  0.1817 -0.1081 -0.0203 -0.0698  0.5531 -0.4655  0.5940
#>  -0.6614  0.4098  0.2428  0.3887 -0.2949 -0.4426  0.7101  0.2345 -0.1185
#> 
#> Columns 19 to 20  0.2067  0.2727
#>  -0.2349 -0.2711
#>   0.0588 -0.1170
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>