RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1949  0.3509  0.1624  0.1003  0.7331 -0.2103  0.2083 -0.3684  0.9458
#>  -0.4603  0.0996  0.2435  0.0203 -0.2642 -0.7532  0.2082  0.1153 -0.4681
#>  -0.4394 -0.5114  0.4715 -0.7136 -0.0019  0.0399  0.7558  0.2387  0.1175
#> 
#> Columns 10 to 18  0.3834 -0.3137  0.1516  0.5919 -0.5516  0.1822  0.7460  0.0371 -0.8701
#>  -0.7936  0.7860 -0.5359 -0.3019 -0.5633 -0.2260 -0.8910 -0.6417  0.9379
#>   0.0743 -0.3975 -0.2425  0.1833  0.9018  0.0364  0.2768  0.1966 -0.7090
#> 
#> Columns 19 to 20 -0.2579 -0.3099
#>  -0.0636 -0.4028
#>   0.3485 -0.7138
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3068  0.2071  0.2207 -0.4223  0.2130 -0.5724 -0.5159 -0.3033  0.2980
#>   0.4920  0.1159  0.3085 -0.3946 -0.5185  0.2091  0.3094  0.5898  0.1757
#>   0.1321 -0.5865  0.4606 -0.2646  0.4666 -0.4117 -0.0635 -0.3413 -0.1572
#> 
#> Columns 10 to 18 -0.1922 -0.1452 -0.3098  0.2915 -0.0002  0.3866  0.3948 -0.2246 -0.1634
#>   0.0970  0.5331  0.2476 -0.0314  0.6979 -0.5393  0.2633  0.1041 -0.1652
#>  -0.0921  0.1856 -0.3801  0.3482 -0.3134  0.3981  0.3030 -0.2500 -0.4900
#> 
#> Columns 19 to 20  0.0747 -0.1636
#>  -0.0543 -0.8117
#>  -0.3098 -0.5752
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1409  0.3581  0.4193 -0.1886 -0.3670 -0.2731 -0.0554 -0.0948  0.3927
#>   0.0470 -0.3428  0.0518  0.6130 -0.1376 -0.3956  0.6891 -0.0317 -0.5187
#>  -0.0429 -0.1681  0.5140 -0.2918 -0.3432 -0.2677 -0.3195  0.0909 -0.0297
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5053  0.1485 -0.2950  0.5952 -0.2301  0.2236  0.2470 -0.1535 -0.0538
#>  -0.0426 -0.2123 -0.8265  0.6002  0.7031  0.3714 -0.1395 -0.3134 -0.1102
#>   0.3650 -0.0628 -0.1985 -0.2490 -0.0568  0.7837  0.6250 -0.3374  0.5738
#> 
#> Columns 10 to 18 -0.1904  0.2268  0.1304 -0.2392  0.6921  0.0695  0.4647 -0.1249  0.0884
#>  -0.7064  0.2322  0.8156 -0.3772  0.1993 -0.0413 -0.4270  0.0230 -0.0131
#>  -0.0422  0.6990  0.8084  0.2123  0.2326  0.0727  0.5933  0.3247  0.8602
#> 
#> Columns 19 to 20 -0.0810 -0.4740
#>  -0.2883  0.0488
#>   0.6108 -0.5560
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3518 -0.4176 -0.0878 -0.2295  0.0763 -0.4544  0.1082  0.2507 -0.1167
#>   0.2593 -0.4638  0.4323 -0.1010 -0.0755 -0.4479  0.0774 -0.2991  0.5004
#>   0.1454  0.0197  0.4067 -0.5749  0.3191 -0.3465 -0.0992  0.1257  0.2061
#> 
#> Columns 10 to 18  0.2072  0.3452  0.1882  0.0701  0.1097  0.0937  0.5072 -0.0148 -0.1287
#>   0.0698 -0.0121 -0.6842 -0.1346 -0.2099  0.2800 -0.0689 -0.4850 -0.0274
#>  -0.1853  0.1896  0.0142  0.0158  0.1426 -0.0288 -0.0639 -0.3880 -0.3201
#> 
#> Columns 19 to 20  0.2243 -0.2698
#>   0.3438 -0.0124
#>  -0.5615 -0.5051
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>