RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1281 -0.7062  0.6662  0.4454  0.6224  0.5567 -0.3867  0.4814  0.1451
#>  -0.0365 -0.9191 -0.0575  0.5781  0.3301 -0.3768  0.4667  0.8210  0.1825
#>   0.1241 -0.1780 -0.4053  0.1696 -0.3064  0.3044  0.3624  0.2148  0.2082
#> 
#> Columns 10 to 18  0.8213 -0.7447 -0.5758 -0.2097  0.1322 -0.2817 -0.5627  0.6545  0.7345
#>  -0.5994  0.7441  0.8857  0.1060 -0.5037  0.4964 -0.1264 -0.5955 -0.6133
#>  -0.3815 -0.7533  0.7928  0.9395  0.8782 -0.4740 -0.4038  0.0639 -0.4938
#> 
#> Columns 19 to 20 -0.2117 -0.1719
#>   0.8775  0.2183
#>  -0.2028 -0.1374
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6207 -0.0195  0.1977  0.3352  0.3722  0.5250 -0.2385  0.0269  0.2263
#>  -0.3819  0.2538  0.2531  0.6291 -0.0194 -0.7039  0.7492  0.0624  0.5031
#>   0.0772 -0.1502  0.1841  0.0551  0.0317  0.3678 -0.3032 -0.5271  0.1956
#> 
#> Columns 10 to 18 -0.2640 -0.3770 -0.2009  0.1878  0.3808 -0.5229  0.4074  0.6908 -0.2748
#>   0.1429  0.4455 -0.3977 -0.5751 -0.4179 -0.7681 -0.4531 -0.3992  0.0941
#>  -0.1958 -0.1531  0.2604 -0.1589  0.0012 -0.0345 -0.2359 -0.0574  0.1799
#> 
#> Columns 19 to 20 -0.1212 -0.6797
#>   0.4405  0.4041
#>  -0.0799  0.0925
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3664 -0.5527  0.0487  0.1901  0.1940 -0.0321  0.2793  0.0408  0.1628
#>  -0.3913 -0.5404  0.0232  0.3153  0.2498 -0.5941  0.2317  0.4756  0.2929
#>  -0.0942 -0.6052 -0.4256  0.1290  0.2458 -0.2239 -0.3111 -0.1194  0.3043
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1291 -0.7940  0.4354 -0.0469 -0.2353 -0.7479  0.5586 -0.0961 -0.0150
#>   0.1905  0.8900 -0.6300  0.1534  0.0442 -0.6430 -0.4145  0.2862  0.5590
#>  -0.4142  0.5647  0.1480 -0.0437 -0.0719  0.2580  0.0640 -0.0766  0.0810
#> 
#> Columns 10 to 18  0.7765 -0.9226 -0.5567  0.0523 -0.5631 -0.3903  0.4378 -0.1522 -0.4348
#>  -0.0193 -0.7417 -0.9002  0.3047  0.3057  0.4164 -0.0538  0.1145  0.6510
#>  -0.1477  0.1641  0.0950 -0.1371 -0.0583  0.4373  0.1414 -0.3166 -0.3309
#> 
#> Columns 19 to 20 -0.0866  0.1077
#>   0.4089 -0.0252
#>   0.0689  0.2463
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4901  0.1103 -0.4252 -0.2176  0.0989  0.1657 -0.2063 -0.7308  0.3893
#>   0.1854 -0.2391 -0.0164 -0.0870  0.1041 -0.6114  0.1240 -0.0696  0.2359
#>   0.5376 -0.1369 -0.2213  0.1678  0.3156  0.3808 -0.0787 -0.0455 -0.0023
#> 
#> Columns 10 to 18 -0.1653 -0.4899 -0.1421  0.3610  0.1356  0.0984 -0.4990  0.6834  0.1285
#>   0.0215  0.1517 -0.4662  0.0485  0.2874 -0.1530 -0.6434  0.0857 -0.0881
#>  -0.2095  0.1828  0.1013  0.1722  0.5287 -0.0658 -0.1603  0.1075 -0.2698
#> 
#> Columns 19 to 20 -0.3974  0.5209
#>  -0.1860  0.2674
#>   0.0258  0.0525
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>