RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9255  0.8142 -0.3455 -0.1318  0.2296 -0.1211  0.8933 -0.4053  0.0605
#>  -0.4190 -0.3928 -0.3684 -0.0815 -0.0202 -0.8734  0.6979  0.0626 -0.2375
#>  -0.2657  0.2943  0.4096  0.0461  0.1357  0.4190  0.0742  0.6606 -0.3055
#> 
#> Columns 10 to 18  0.8770 -0.9094 -0.1868  0.2396  0.4359  0.6258 -0.0279 -0.2155  0.9408
#>   0.0097 -0.3642 -0.1188 -0.5725  0.2177 -0.6006 -0.2873 -0.7482 -0.0440
#>  -0.3868  0.2661 -0.7791 -0.1350  0.3472  0.5558  0.4322  0.6276  0.1233
#> 
#> Columns 19 to 20 -0.9459 -0.5039
#>   0.1471 -0.4973
#>   0.0857  0.6927
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0492  0.3623 -0.1597 -0.7229  0.4304  0.3171  0.5646 -0.3230  0.1646
#>  -0.3451 -0.1273  0.1785 -0.6053 -0.1259  0.5171 -0.0681 -0.1099 -0.2208
#>  -0.1635 -0.2079  0.2340  0.2693  0.5356  0.0915  0.3832  0.4442 -0.4351
#> 
#> Columns 10 to 18  0.1190  0.5414 -0.4164  0.3390 -0.0171  0.0842  0.2238  0.5915 -0.0061
#>   0.2588 -0.0544  0.5066  0.4118 -0.6409 -0.3234 -0.0368  0.5474 -0.0633
#>   0.4109  0.0844 -0.1986 -0.2279  0.2921  0.1380  0.4558 -0.1683  0.6284
#> 
#> Columns 19 to 20 -0.1201  0.2191
#>   0.2871 -0.0739
#>  -0.2522  0.6043
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0634 -0.1029 -0.1651 -0.1419  0.5310 -0.0045  0.3081  0.2731 -0.1399
#>  -0.0997  0.2884  0.2678 -0.6843  0.4957  0.3576  0.6543  0.3731 -0.0853
#>  -0.2193  0.3217  0.1348 -0.4055  0.4685 -0.2920  0.3580 -0.1913 -0.3081
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1337 -0.2477 -0.4285 -0.3906  0.3763  0.1869 -0.5528 -0.6327  0.0360
#>  -0.1823  0.1957 -0.5362 -0.8458 -0.3648 -0.1886 -0.5851  0.2742 -0.6197
#>   0.3923  0.3293  0.4205  0.2226  0.5639  0.0210 -0.5290 -0.0695 -0.2170
#> 
#> Columns 10 to 18  0.1454  0.3663 -0.6722 -0.4915  0.0770 -0.1965 -0.2371  0.1568 -0.6373
#>   0.7606  0.5649 -0.6599  0.1869  0.0306  0.2787 -0.1328  0.1632  0.4845
#>  -0.1449 -0.2381 -0.4472  0.3683  0.1623 -0.0807 -0.5066  0.1685  0.4439
#> 
#> Columns 19 to 20 -0.4155 -0.2237
#>   0.3808  0.4577
#>   0.0064  0.0759
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1745  0.0333  0.5018 -0.6050  0.0851  0.1618  0.2099  0.0522 -0.0428
#>  -0.2037  0.5893 -0.3807 -0.3823  0.3935 -0.1317  0.3837  0.1046  0.1137
#>   0.4138  0.1342  0.3414 -0.1325  0.5329 -0.2594  0.2225 -0.3672  0.0614
#> 
#> Columns 10 to 18  0.7450  0.5007  0.1843  0.2557 -0.4916  0.0329 -0.1807  0.3969  0.4174
#>   0.5331  0.3057  0.1476  0.0266 -0.3232  0.1896 -0.5212 -0.1719  0.1171
#>   0.5889  0.0017  0.0676  0.3826  0.1621 -0.6203  0.0731  0.2412  0.1106
#> 
#> Columns 19 to 20  0.4414  0.1144
#>  -0.2406  0.0232
#>   0.6585 -0.0592
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>