RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3754 -0.5915 -0.7473 -0.3055 -0.8013  0.8829  0.0535 -0.7560 -0.1124
#>   0.3363  0.0026  0.2634  0.2838  0.7081  0.5116  0.7131 -0.2074  0.3113
#>   0.0611 -0.3401  0.7844  0.2927  0.1679  0.6810 -0.1686  0.5864 -0.8313
#> 
#> Columns 10 to 18  0.7390 -0.1469  0.3434  0.5975  0.6632 -0.6300 -0.5835  0.7972  0.1096
#>   0.0009  0.3037  0.4651 -0.2802 -0.3759 -0.7812 -0.1761  0.7363  0.2635
#>   0.2880 -0.0894 -0.7535 -0.2826  0.7229 -0.6311 -0.5622 -0.4072  0.1579
#> 
#> Columns 19 to 20 -0.7323  0.2038
#>  -0.4943  0.2823
#>   0.5806 -0.2582
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0546 -0.3645  0.7486  0.4080 -0.6913 -0.1090  0.1781  0.2298 -0.0766
#>   0.7925 -0.3319  0.3239  0.5582 -0.2094 -0.2399 -0.3979  0.0659  0.0332
#>   0.5175 -0.3484  0.8282  0.0535  0.1651  0.4135 -0.4045  0.3552 -0.0948
#> 
#> Columns 10 to 18 -0.3815 -0.3823 -0.2234 -0.3897 -0.3893  0.3074 -0.6582 -0.3431  0.2219
#>   0.4815 -0.3192 -0.0198 -0.0440 -0.1689  0.4724 -0.6650 -0.1559  0.6515
#>  -0.0959 -0.3674 -0.0939 -0.3642 -0.2114 -0.3731 -0.6321 -0.0050  0.6537
#> 
#> Columns 19 to 20 -0.0571 -0.0937
#>   0.4262  0.1013
#>   0.1057 -0.3910
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2897 -0.3110  0.2688 -0.2389  0.0696  0.4854  0.3040 -0.2415 -0.0661
#>   0.4850 -0.0841  0.6050 -0.3327 -0.1197  0.1294  0.2505 -0.5092  0.4525
#>   0.5726 -0.5065  0.6882 -0.2791 -0.1595  0.1333 -0.5570 -0.2736 -0.2322
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7678  0.4442 -0.4927 -0.4658 -0.1624 -0.4228 -0.4822  0.5146 -0.4442
#>  -0.1748 -0.1827 -0.3342  0.1738  0.5187  0.3560  0.4685 -0.4394 -0.7089
#>  -0.2476  0.1123 -0.4431  0.2302  0.3699 -0.0138 -0.0864 -0.2202 -0.3695
#> 
#> Columns 10 to 18  0.1197 -0.5273  0.5284 -0.1810  0.0052  0.7475 -0.0543 -0.1951  0.7593
#>  -0.7250  0.4750  0.1626 -0.7279  0.7731 -0.3551  0.5990  0.5111  0.4220
#>  -0.3149  0.6167  0.0327  0.2279  0.3869 -0.3156  0.2096  0.2563 -0.1763
#> 
#> Columns 19 to 20  0.1051 -0.4035
#>  -0.3486  0.4935
#>  -0.8971  0.6415
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2477 -0.5540  0.7003  0.0089 -0.1679  0.2223  0.4501  0.0258  0.2882
#>   0.3086 -0.4683  0.7472 -0.1606 -0.2994 -0.0357  0.0707 -0.3865 -0.0983
#>   0.3769 -0.4494  0.5405 -0.2851  0.0045 -0.1436  0.2068 -0.2201  0.0253
#> 
#> Columns 10 to 18  0.1853 -0.3298 -0.2191 -0.5419 -0.1313 -0.2181 -0.4724  0.0517  0.1849
#>   0.4359 -0.4087 -0.0077 -0.2773 -0.0434 -0.1559 -0.8020  0.2672  0.3775
#>   0.6329 -0.3305  0.0442 -0.4053 -0.0170  0.2102 -0.8028  0.2367  0.4488
#> 
#> Columns 19 to 20 -0.1576 -0.3600
#>   0.1682 -0.4026
#>  -0.0699 -0.3699
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>