RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0591  0.6665  0.1429  0.6731 -0.0063 -0.6505  0.1417  0.6382  0.1157
#>  -0.3082 -0.1167 -0.5762 -0.8491 -0.2294 -0.0778 -0.2903 -0.2353  0.4172
#>  -0.7439  0.5813 -0.2726 -0.7740 -0.8109 -0.6382  0.0831 -0.7481 -0.4172
#> 
#> Columns 10 to 18 -0.1962 -0.3270 -0.8474  0.2148 -0.8466  0.3675  0.4547 -0.3917  0.7196
#>   0.8872  0.0784  0.5593  0.2294  0.3411 -0.2967 -0.2167 -0.2717 -0.0007
#>   0.0215  0.3116  0.0517  0.8076  0.2301 -0.0703 -0.4024  0.2550  0.3576
#> 
#> Columns 19 to 20  0.3987 -0.5714
#>   0.7845 -0.2313
#>  -0.7860  0.2282
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4942 -0.1959 -0.1405  0.1172 -0.4171  0.6194 -0.1902  0.1587  0.3133
#>   0.0108  0.4770  0.4642  0.0345 -0.2446 -0.0366  0.1133 -0.0935 -0.1416
#>  -0.0715 -0.3862 -0.1016  0.3097  0.0653  0.1734  0.4762  0.2537 -0.7420
#> 
#> Columns 10 to 18 -0.4409  0.0036 -0.2600  0.2089  0.2932  0.0150 -0.0930  0.0223  0.2478
#>  -0.1053  0.1063  0.2271 -0.1283 -0.5599 -0.1430  0.1017  0.3481  0.1655
#>  -0.3298 -0.2017 -0.2247  0.3388 -0.6000 -0.5419 -0.2668  0.0918  0.2537
#> 
#> Columns 19 to 20  0.3360  0.0408
#>  -0.1761 -0.5587
#>  -0.2625 -0.2700
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1307 -0.0408  0.1977  0.3083  0.3050 -0.1084 -0.2358 -0.0248  0.0615
#>   0.1499  0.0105  0.2987  0.2769 -0.0066 -0.0029 -0.3705  0.2758 -0.2938
#>   0.1793  0.0307 -0.1751  0.3609 -0.1563 -0.0461 -0.3585 -0.3478 -0.3367
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4335 -0.0665  0.5889 -0.2627  0.0950  0.2210  0.2669  0.5380  0.3431
#>  -0.3924  0.2022 -0.4507 -0.3655 -0.0261 -0.4941 -0.0127  0.0046 -0.1571
#>   0.3218  0.1374 -0.1057 -0.0101 -0.1038  0.2516 -0.1129 -0.2679  0.1140
#> 
#> Columns 10 to 18  0.1303  0.1444 -0.5151 -0.3488  0.2336 -0.6537 -0.3875 -0.3520 -0.3264
#>   0.1932 -0.3217  0.0088  0.2078 -0.5479 -0.5539  0.3327 -0.5904  0.3707
#>   0.1310  0.3056  0.0911  0.3473  0.1577 -0.2062  0.1816 -0.5300 -0.4125
#> 
#> Columns 19 to 20 -0.0541 -0.4752
#>   0.0136 -0.0188
#>   0.7931 -0.3827
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0975  0.1411  0.0698  0.4174 -0.6399  0.2341 -0.3532  0.0240 -0.1054
#>   0.2316 -0.1103  0.2398  0.2881 -0.1449 -0.3220 -0.0861  0.0211 -0.2893
#>   0.1971  0.3510  0.2129  0.2347 -0.6301  0.0541 -0.4308  0.0975 -0.0275
#> 
#> Columns 10 to 18 -0.4968  0.0547 -0.4418  0.2308 -0.1585  0.0406 -0.2860  0.4319  0.4591
#>  -0.4552  0.0749 -0.4093  0.0998 -0.3547  0.0689  0.3840 -0.2815  0.5139
#>  -0.6175  0.0735 -0.5272  0.3755  0.0892  0.0397 -0.1768 -0.3333  0.0356
#> 
#> Columns 19 to 20 -0.1984 -0.1898
#>  -0.0902 -0.2562
#>   0.1157 -0.1716
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>