RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6181  0.0930  0.1997  0.4828 -0.0576 -0.6934 -0.2512  0.4466 -0.1718
#>  -0.6241  0.7238  0.5827  0.4565  0.3208  0.0128 -0.6842  0.7012  0.1590
#>  -0.8680  0.7532  0.1442 -0.1110  0.1103 -0.7391 -0.6635  0.4344  0.7111
#> 
#> Columns 10 to 18 -0.6751 -0.7259  0.6056  0.7118  0.1393  0.1132 -0.8079  0.5736  0.5221
#>   0.2405  0.0720  0.1035 -0.0029 -0.0830 -0.1961  0.1431 -0.5438 -0.2801
#>  -0.0786 -0.2886  0.1929  0.0811  0.7439 -0.4928 -0.8261 -0.0431  0.8602
#> 
#> Columns 19 to 20  0.2257 -0.3313
#>  -0.1615 -0.6571
#>   0.3764 -0.3764
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5995  0.2905 -0.3057  0.4464 -0.1514 -0.1323  0.0460  0.3910 -0.7817
#>   0.1886  0.3314 -0.2472  0.4843 -0.0164  0.4717  0.3465  0.3772 -0.3604
#>   0.1572 -0.2290 -0.0906  0.6414 -0.3524  0.2095  0.4578  0.7476 -0.7989
#> 
#> Columns 10 to 18  0.6265 -0.0534 -0.2207 -0.1710  0.4071 -0.6456 -0.7040 -0.3353  0.6383
#>   0.2464 -0.5709 -0.1888 -0.7051 -0.2481 -0.4277 -0.5703 -0.2882  0.0029
#>   0.3261 -0.2276 -0.4741 -0.2698  0.2616 -0.4189 -0.1369 -0.3310 -0.1064
#> 
#> Columns 19 to 20 -0.4073 -0.2153
#>  -0.0637 -0.3593
#>  -0.1656  0.0878
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.6206 -0.4044 -0.1276  0.6784  0.0037  0.4241  0.3214  0.5162 -0.2124
#>   0.2079 -0.5710 -0.5607  0.2089 -0.1703 -0.0149 -0.3292 -0.2222 -0.2167
#>  -0.1708 -0.6562 -0.3285  0.4832  0.0608 -0.5462 -0.2236 -0.0877  0.6182
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6192 -0.1944 -0.3171 -0.3090 -0.5342 -0.0771  0.3877 -0.3424  0.3474
#>   0.2132 -0.5743  0.0309 -0.2006  0.1292  0.1052  0.2252 -0.1106  0.1578
#>  -0.1036  0.4811 -0.2163 -0.0695 -0.0187  0.1013  0.3350  0.4153  0.7513
#> 
#> Columns 10 to 18 -0.5675  0.3050 -0.2371  0.5894 -0.0927 -0.1418 -0.0307  0.5289  0.1104
#>  -0.2451 -0.0837  0.4784 -0.4998  0.0370 -0.3785 -0.1572  0.7495  0.0298
#>   0.0200 -0.3572  0.6501 -0.8247 -0.4832 -0.3945  0.3908  0.1420 -0.5697
#> 
#> Columns 19 to 20  0.1215 -0.1822
#>  -0.4191 -0.5719
#>   0.0213  0.1372
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1787  0.0322  0.3771  0.1595  0.0856 -0.3013  0.3742  0.3535  0.0501
#>  -0.0153  0.0156  0.1625  0.3776  0.1390 -0.0292  0.3170  0.3066 -0.0087
#>   0.3447 -0.1377 -0.1374  0.1262 -0.0501  0.2996  0.2391  0.0254  0.2443
#> 
#> Columns 10 to 18 -0.4910 -0.4592 -0.5642  0.4482  0.0038  0.2247 -0.4293 -0.1931  0.0349
#>  -0.1956 -0.3384 -0.1363 -0.2247  0.2777 -0.3091 -0.2116  0.0100  0.4379
#>  -0.6447 -0.3500  0.1520 -0.0109  0.1462 -0.3003 -0.0936 -0.3872  0.4059
#> 
#> Columns 19 to 20 -0.0223 -0.1270
#>   0.4583  0.1013
#>   0.4921 -0.1020
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>