RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5824 -0.8068  0.1111  0.5699 -0.1929  0.4524 -0.7724 -0.5246 -0.9348
#>   0.0649  0.5791 -0.5894  0.1696  0.1501  0.4367 -0.0500  0.9450  0.3684
#>  -0.7833  0.2117  0.8493  0.2744  0.7655  0.6264 -0.0568  0.1859  0.7320
#> 
#> Columns 10 to 18  0.5719  0.1646  0.5135  0.7688  0.3537  0.1449 -0.6149  0.8751 -0.1612
#>  -0.6438 -0.2687  0.5564 -0.4635 -0.8609 -0.6570  0.7660 -0.1383  0.7251
#>  -0.7296 -0.5170  0.2434  0.4210 -0.5458  0.2913  0.5504  0.7368  0.4720
#> 
#> Columns 19 to 20  0.0662  0.8603
#>  -0.8531  0.5346
#>  -0.3377  0.2689
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0856 -0.3722 -0.0748 -0.3384  0.2967  0.3974  0.0897  0.5339  0.0740
#>  -0.3236  0.0725  0.3047 -0.1748 -0.0348  0.3723 -0.2032  0.2754 -0.4933
#>   0.2432 -0.1318  0.6923  0.2234  0.1835  0.0524  0.3116  0.1660  0.0894
#> 
#> Columns 10 to 18  0.2707 -0.2902 -0.4544 -0.2321  0.1142 -0.4672  0.5279  0.2721  0.1202
#>   0.1911 -0.0584  0.5338  0.4182  0.1439 -0.2373 -0.7012  0.1777  0.2069
#>   0.3816  0.0820  0.5310 -0.0233  0.3713  0.0177 -0.2666 -0.3178  0.1451
#> 
#> Columns 19 to 20 -0.2609  0.3837
#>   0.6152  0.4384
#>  -0.4609  0.1123
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2797  0.3644  0.2239 -0.2958  0.2729  0.6087  0.2071  0.4009  0.4054
#>  -0.1848 -0.2123  0.4263 -0.2432  0.3963  0.4202 -0.0324  0.2963  0.1412
#>   0.3231  0.0697  0.3076 -0.3396  0.3496 -0.0817 -0.2378  0.2084 -0.5872
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8355 -0.8368 -0.1599  0.6608  0.2643 -0.3588  0.5062  0.4307 -0.2887
#>  -0.3748  0.6348  0.5819  0.6555 -0.5016 -0.3139 -0.1122  0.1281 -0.1572
#>  -0.7250 -0.5795 -0.1176  0.5396  0.4369 -0.1550 -0.4146  0.6763 -0.7422
#> 
#> Columns 10 to 18  0.6539  0.8748  0.2211  0.1805 -0.8415  0.6194 -0.3964  0.4753 -0.5413
#>  -0.2174 -0.1283 -0.2642 -0.4738 -0.0442 -0.1722 -0.4991 -0.3473  0.6837
#>  -0.4443  0.6267  0.2211  0.2286 -0.2390 -0.0153 -0.6583  0.4078 -0.4936
#> 
#> Columns 19 to 20 -0.2253 -0.2754
#>   0.2357  0.1336
#>   0.4055  0.4319
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5582 -0.4154  0.3279 -0.0819  0.3488 -0.4832  0.0651 -0.4595 -0.1465
#>  -0.1139  0.3960  0.4028 -0.0983  0.1753 -0.2354 -0.4362  0.2595 -0.6890
#>  -0.5433 -0.4583  0.4301 -0.2810  0.6409  0.3219 -0.1766 -0.1998  0.3940
#> 
#> Columns 10 to 18  0.2598 -0.1205 -0.1088  0.7277  0.3738  0.2021  0.0427 -0.2386  0.2377
#>  -0.3076 -0.2548  0.4125 -0.0602  0.5270  0.0320 -0.3460 -0.3025  0.3641
#>  -0.2589  0.0348 -0.0707  0.4917 -0.0789 -0.4263  0.1479  0.1624  0.1380
#> 
#> Columns 19 to 20 -0.2819  0.0537
#>   0.2464  0.2615
#>  -0.3871  0.3372
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>