RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8878 -0.6330 -0.3457 -0.1010 -0.6719  0.8040  0.0921  0.0072  0.8756
#>   0.1350  0.0803  0.7756 -0.7725 -0.6702 -0.0448  0.2813 -0.9537 -0.4294
#>  -0.5591 -0.5759 -0.4872 -0.0738  0.3443 -0.4641 -0.6602 -0.6872 -0.3681
#> 
#> Columns 10 to 18 -0.3844  0.2456 -0.8737  0.0073  0.5739 -0.1222 -0.6739  0.1756  0.3705
#>  -0.0286  0.4587 -0.6955 -0.0911  0.6880 -0.5164 -0.1114  0.7252 -0.0201
#>   0.2177 -0.6543 -0.7867 -0.2309  0.2286  0.3569 -0.3233  0.0460  0.0728
#> 
#> Columns 19 to 20  0.1274  0.0773
#>  -0.4903  0.7423
#>  -0.3098 -0.6441
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0986  0.1175 -0.0580 -0.2899 -0.1134  0.2684  0.2556  0.2914  0.2287
#>   0.7217  0.0276  0.0665 -0.5041 -0.6140  0.5232  0.1301  0.3953  0.2376
#>   0.1825  0.0502  0.1830 -0.4975 -0.0459 -0.2399  0.3665 -0.5169  0.0076
#> 
#> Columns 10 to 18  0.1415  0.0505  0.4748  0.0017 -0.6798 -0.1504  0.3576  0.3537  0.3364
#>  -0.0097  0.3328  0.1217 -0.4189 -0.5976 -0.1136  0.1162 -0.2756 -0.2381
#>   0.0415 -0.0408 -0.1033  0.5790  0.3370  0.0637 -0.0952  0.0252  0.0349
#> 
#> Columns 19 to 20 -0.0012 -0.3326
#>   0.0530  0.0270
#>  -0.7211 -0.4137
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0057 -0.3492 -0.0619 -0.3376  0.0544  0.4955  0.3883 -0.4373  0.6142
#>  -0.0303 -0.2604 -0.2755  0.0351 -0.3439  0.6194 -0.0575 -0.1376  0.4492
#>   0.4791  0.0330  0.0478 -0.3266 -0.1829  0.0361  0.0099  0.1884 -0.0744
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0765 -0.3822 -0.4714  0.3985 -0.0846  0.7645  0.4259 -0.0368 -0.2476
#>   0.3548  0.2862  0.2864  0.1063 -0.1565  0.5628  0.2712  0.5151  0.2674
#>   0.6377 -0.1107  0.2524  0.4153 -0.1015  0.1650  0.6925  0.6359 -0.1720
#> 
#> Columns 10 to 18  0.5487  0.4981 -0.0339  0.1742  0.7422  0.7072  0.1120  0.5192 -0.3730
#>  -0.0923  0.4555  0.2590  0.1610 -0.1168 -0.0343  0.6183 -0.0573  0.3392
#>   0.1948 -0.3259 -0.2597  0.5729  0.1282  0.6644  0.4558  0.3785  0.6117
#> 
#> Columns 19 to 20  0.3530 -0.0334
#>  -0.0307  0.1563
#>  -0.3200 -0.0891
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1972 -0.1294 -0.1859 -0.1759 -0.0288  0.5391  0.5977 -0.3831  0.3630
#>  -0.0978 -0.3600 -0.2337 -0.1876 -0.0057  0.1420  0.3603 -0.2330  0.0742
#>   0.3765 -0.2617 -0.1228 -0.2577 -0.2028  0.1981  0.2574 -0.1092 -0.0356
#> 
#> Columns 10 to 18 -0.2122 -0.1095 -0.2466 -0.1834 -0.2481 -0.2155 -0.1641  0.1719  0.3141
#>   0.2125  0.2687 -0.3245 -0.3910 -0.3743 -0.2894  0.0568  0.2013 -0.1456
#>  -0.0720  0.1737 -0.3210 -0.3654 -0.5643 -0.0238  0.0308 -0.3906 -0.6165
#> 
#> Columns 19 to 20  0.1233 -0.4189
#>   0.1328 -0.3190
#>  -0.2838 -0.2675
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>