RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4884 -0.2650  0.1179 -0.3494 -0.5611 -0.5708  0.8002  0.6569 -0.3654
#>  -0.1437 -0.6269 -0.0801 -0.0186 -0.4452 -0.5909 -0.7469 -0.4099 -0.3591
#>   0.6850  0.5826  0.5636  0.7950  0.6936  0.1232  0.4041 -0.1672 -0.2508
#> 
#> Columns 10 to 18 -0.0081  0.4408  0.2299 -0.3854  0.0201  0.1578 -0.8579  0.7536  0.3373
#>   0.1372  0.0350 -0.5956 -0.0740  0.3580 -0.1829  0.6525 -0.4393  0.3574
#>  -0.7785  0.9123 -0.6589  0.6958 -0.6921 -0.3094  0.4016  0.0358  0.8373
#> 
#> Columns 19 to 20  0.4685 -0.4921
#>  -0.4378  0.3884
#>  -0.1085 -0.1217
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7929 -0.4527  0.5220  0.2990  0.1222  0.2346  0.4049 -0.2806 -0.7879
#>   0.4436 -0.0371  0.2590  0.3648  0.3643 -0.5612  0.0068 -0.3038 -0.1624
#>  -0.1383 -0.4108 -0.4047  0.0535 -0.3378  0.0964  0.3204  0.6765 -0.5412
#> 
#> Columns 10 to 18 -0.1781  0.2603  0.3140  0.2598  0.1522 -0.3651  0.2486 -0.0929  0.0535
#>  -0.1596  0.4966  0.1317  0.0354 -0.4128 -0.1084 -0.3073  0.6454  0.5268
#>   0.0956  0.0763  0.1488  0.5112 -0.4094  0.1843 -0.3974  0.6607  0.2086
#> 
#> Columns 19 to 20 -0.4244 -0.2325
#>  -0.0556 -0.1986
#>   0.1962 -0.4299
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3740  0.5333  0.3787  0.3700 -0.0021  0.0841 -0.2217 -0.2376 -0.3407
#>   0.0172  0.2593  0.4494 -0.2535  0.1571 -0.2959 -0.1144 -0.2513  0.0054
#>  -0.3561 -0.1225  0.5445 -0.0251  0.0094 -0.0675  0.1020 -0.4932 -0.5463
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7473 -0.4354  0.3902 -0.2594 -0.0334 -0.2715  0.5222  0.4081 -0.1428
#>   0.3847  0.2777 -0.5949  0.1407  0.1006 -0.4908  0.6071  0.4152  0.0314
#>   0.4781  0.4745 -0.0791  0.4953  0.3885 -0.1369 -0.0173  0.3711  0.0826
#> 
#> Columns 10 to 18  0.0275 -0.3815  0.1704  0.4051 -0.3107 -0.4268  0.8153  0.4852 -0.4853
#>   0.7309 -0.1764 -0.3406 -0.1792  0.3868 -0.6093 -0.6621  0.2505 -0.1346
#>   0.1984 -0.1661 -0.0932 -0.1195 -0.2734  0.2098  0.2050  0.5714  0.2755
#> 
#> Columns 19 to 20  0.9344  0.8372
#>  -0.7267 -0.1116
#>  -0.0297 -0.1035
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5888 -0.1191  0.2222 -0.0186 -0.0809  0.0130  0.0541 -0.4928 -0.5453
#>  -0.2028  0.3846  0.4537 -0.1548  0.5660 -0.2805 -0.1718 -0.2191 -0.3935
#>  -0.2364  0.2758  0.2845  0.5074  0.3291 -0.2272 -0.1333 -0.0584 -0.5375
#> 
#> Columns 10 to 18  0.0152 -0.3813  0.8139 -0.1899  0.3355 -0.0259 -0.0011  0.0826  0.2959
#>   0.3039  0.0651  0.2083 -0.6127 -0.0247  0.2663  0.2557  0.2778  0.5180
#>   0.3008  0.0934  0.1278 -0.1657  0.0955  0.0873  0.1799  0.4189  0.5878
#> 
#> Columns 19 to 20 -0.0619 -0.2946
#>  -0.2613 -0.0738
#>   0.0103 -0.1486
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>