RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4567  0.2942 -0.3100  0.0219  0.4944 -0.5798  0.8862  0.3544 -0.3223
#>   0.2061 -0.1877  0.3433  0.0810  0.4155  0.4283  0.6729 -0.0501 -0.4122
#>  -0.5809  0.1104  0.1213  0.4539  0.2699  0.5292  0.7807  0.0318 -0.2503
#> 
#> Columns 10 to 18  0.3439  0.9044 -0.7803 -0.6465  0.4554 -0.2440  0.4172  0.6905  0.0613
#>  -0.3672  0.7102  0.1340 -0.4021 -0.1261 -0.2815  0.1606 -0.0097 -0.6028
#>  -0.1686 -0.7849 -0.7140 -0.5652  0.1666 -0.6083 -0.6058  0.3727  0.4228
#> 
#> Columns 19 to 20  0.3024 -0.3180
#>  -0.1383 -0.2920
#>   0.4452 -0.7193
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0323  0.0775  0.1976 -0.7824 -0.3752  0.1710  0.2174  0.3596 -0.7077
#>   0.0059 -0.1258 -0.3468 -0.5742  0.1625  0.2674 -0.2162 -0.0415 -0.4238
#>  -0.3919  0.5417  0.2031  0.0197 -0.0787  0.4967  0.4350  0.7527 -0.7411
#> 
#> Columns 10 to 18  0.4003 -0.2350 -0.4126  0.2988  0.6084  0.3118 -0.7840  0.6528 -0.1482
#>  -0.0860  0.3766 -0.4767  0.1180  0.1991 -0.0167 -0.1878 -0.3199 -0.4829
#>  -0.3643 -0.2437 -0.6261 -0.3221  0.6039 -0.5062 -0.7664  0.2241 -0.2303
#> 
#> Columns 19 to 20  0.1849 -0.3613
#>   0.5559 -0.5757
#>   0.1356  0.0263
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5025 -0.0017 -0.0587 -0.1455  0.4854  0.6568  0.5519  0.1869 -0.3557
#>  -0.0159 -0.2274 -0.3302  0.0816  0.3708  0.6084  0.0725  0.4568 -0.5668
#>  -0.4452 -0.0362  0.1439 -0.3095 -0.5074  0.4925  0.5288  0.6762 -0.6004
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0868 -0.5809 -0.1644 -0.2710 -0.7563 -0.2976  0.0180 -0.5789 -0.4019
#>  -0.2198 -0.3504  0.6014  0.1920 -0.5402 -0.4228 -0.4721 -0.3091 -0.4353
#>  -0.1060  0.3347  0.5049 -0.2723 -0.3241  0.1163 -0.4951 -0.1146 -0.4655
#> 
#> Columns 10 to 18 -0.3194 -0.2775 -0.5710  0.0420 -0.4267  0.6363  0.7224  0.6689 -0.2867
#>   0.2641 -0.1713  0.3784 -0.5289  0.4966 -0.1352 -0.1949 -0.2314 -0.5791
#>  -0.1043  0.3713 -0.3078 -0.6269  0.1347 -0.1985 -0.5897 -0.1751 -0.2522
#> 
#> Columns 19 to 20  0.4545 -0.5264
#>   0.3297  0.3789
#>   0.0615  0.2016
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1139  0.0832  0.1015 -0.1760  0.4325  0.6058 -0.1224  0.5818 -0.5176
#>  -0.1514 -0.4398  0.2215 -0.6536 -0.1567  0.5634  0.3347  0.6399 -0.6595
#>  -0.2154 -0.1002  0.0423 -0.3300 -0.0629  0.4945  0.4210  0.6601 -0.7763
#> 
#> Columns 10 to 18 -0.1216  0.1672 -0.7214 -0.5136  0.3756 -0.1974 -0.4374  0.0189 -0.2368
#>  -0.0390  0.3857 -0.5479 -0.1398  0.7544  0.0185 -0.6116  0.1744 -0.2038
#>  -0.0001 -0.0839 -0.7480 -0.2703  0.8188  0.0371 -0.6427  0.5130  0.0324
#> 
#> Columns 19 to 20  0.6306 -0.4335
#>   0.3259 -0.4675
#>   0.1706 -0.4703
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>