RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2923 -0.2822  0.8516 -0.7299 -0.3959  0.4350 -0.0615 -0.6068  0.9651
#>   0.4113  0.1357 -0.2198  0.5691  0.1120  0.6405 -0.4494  0.1421 -0.1649
#>   0.4905 -0.8907  0.4385 -0.2453  0.2583 -0.6568  0.1657 -0.4323  0.9263
#> 
#> Columns 10 to 18 -0.3169 -0.1933  0.6894  0.1699 -0.5964  0.1036  0.9042  0.1950  0.3466
#>   0.6244  0.7217 -0.3989  0.2467  0.7752  0.4598 -0.1403  0.7578  0.6143
#>  -0.2815 -0.1811  0.8923  0.3150 -0.5009  0.5215  0.6063  0.0358 -0.2759
#> 
#> Columns 19 to 20 -0.0689 -0.3429
#>   0.2059 -0.4980
#>   0.5580  0.1886
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2473  0.4873  0.0419  0.2879  0.8342  0.3325 -0.2908 -0.1577  0.2079
#>   0.5027  0.0923  0.5295  0.1919  0.3778  0.6081 -0.0917  0.0358 -0.3142
#>   0.2967 -0.4086 -0.2294  0.0376  0.4318 -0.1689  0.0621  0.0639  0.5442
#> 
#> Columns 10 to 18  0.2432  0.6452 -0.6297 -0.0627  0.1820 -0.2592 -0.0733 -0.3634  0.4742
#>   0.1407 -0.0248  0.1103  0.3743 -0.0491  0.2295  0.3603  0.4462  0.0022
#>   0.1259  0.1563  0.0479 -0.1298  0.6336  0.0654 -0.1493  0.1945  0.8094
#> 
#> Columns 19 to 20 -0.1473 -0.4867
#>   0.1388  0.2770
#>   0.6663 -0.5373
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5006  0.1780 -0.0976 -0.0943  0.5487  0.5013  0.1735  0.1975 -0.0448
#>  -0.1441 -0.1163  0.2169 -0.6595  0.3497  0.5207  0.1884  0.4379 -0.0210
#>   0.2173 -0.0046  0.0117  0.3436  0.4435  0.1466  0.0158  0.1369  0.4968
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2459 -0.5259 -0.2916 -0.1806  0.1273 -0.0683 -0.1798 -0.2746 -0.4579
#>  -0.0170  0.4274  0.3634 -0.5873  0.1836  0.1906 -0.4405 -0.7029  0.2374
#>   0.4681 -0.1479  0.0463 -0.3480 -0.0840  0.0375 -0.5612  0.0866 -0.6408
#> 
#> Columns 10 to 18 -0.0762  0.0345  0.6361 -0.4171 -0.0539  0.0976 -0.0699  0.3860  0.2985
#>  -0.7322 -0.3043  0.7422 -0.2897  0.5184 -0.7179  0.3554 -0.7609  0.0702
#>  -0.1871  0.6916  0.6435  0.0869  0.0238  0.2341  0.1443 -0.2863  0.2648
#> 
#> Columns 19 to 20  0.1169  0.3950
#>   0.2703 -0.3401
#>   0.6755  0.3645
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4873  0.0424  0.2711  0.1877  0.5905  0.4760 -0.0272  0.0072 -0.2678
#>   0.3753  0.0703  0.2377 -0.1179  0.1998 -0.1841  0.1891 -0.3861 -0.0438
#>   0.0821 -0.0050  0.3835  0.2074  0.6589  0.4561 -0.0950  0.0673  0.2652
#> 
#> Columns 10 to 18 -0.1003  0.2615 -0.2171  0.2529  0.0317  0.2285  0.0577  0.0556  0.0968
#>  -0.4628  0.5215  0.4203  0.2772 -0.5015  0.1512  0.5931 -0.3467  0.2538
#>   0.0681  0.1233 -0.0439  0.6487  0.1721  0.1866  0.1801 -0.0900  0.3624
#> 
#> Columns 19 to 20  0.1339 -0.2480
#>   0.2845  0.0288
#>   0.1831 -0.0261
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>