RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7443 -0.4755  0.5374 -0.4184 -0.8156  0.8809  0.0453  0.2960 -0.7068
#>  -0.6059  0.6491  0.6696 -0.2186 -0.7122 -0.2774  0.3248 -0.8637 -0.2982
#>   0.7541  0.1775 -0.6313  0.6375  0.6820  0.0582  0.1762 -0.2266 -0.7457
#> 
#> Columns 10 to 18  0.4067  0.8844  0.5704 -0.3557  0.2724  0.0017 -0.6032  0.3744 -0.0156
#>  -0.7283  0.9025  0.8750 -0.1727 -0.2637 -0.1175  0.0924  0.9060  0.0108
#>  -0.5519  0.1109  0.3640  0.0079  0.7046  0.6097 -0.0455 -0.2788  0.0866
#> 
#> Columns 19 to 20 -0.4725 -0.9656
#>  -0.9724 -0.0076
#>   0.4785 -0.0687
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2906 -0.1066  0.6920 -0.4451 -0.0721  0.2398  0.1175 -0.0628 -0.1062
#>   0.5528  0.1685 -0.2593 -0.5729 -0.0159  0.5018 -0.4070  0.4706 -0.1890
#>  -0.1119  0.0867  0.2516  0.3856 -0.0653  0.0734  0.1371 -0.4333  0.0626
#> 
#> Columns 10 to 18  0.2398  0.0530 -0.3113 -0.4410 -0.5687  0.6854  0.3472  0.0505  0.4550
#>  -0.5460  0.5716  0.4040  0.1068  0.0006  0.3588  0.0075  0.1913  0.3230
#>  -0.7415  0.0711  0.1119  0.2388  0.0569  0.6136  0.0637  0.4995  0.2013
#> 
#> Columns 19 to 20  0.1986 -0.4598
#>   0.1108 -0.3402
#>  -0.1284  0.2696
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3498  0.6189  0.6230 -0.3018  0.4406  0.5545 -0.0928 -0.1346 -0.6151
#>   0.4913  0.0731  0.1986 -0.0142  0.3147 -0.0535 -0.5124  0.0785 -0.6658
#>   0.5312  0.1899 -0.3683 -0.0937  0.2614 -0.0837 -0.2873 -0.1079 -0.4172
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1483 -0.3994 -0.2699 -0.2886  0.1017  0.1457 -0.0994  0.1356  0.1036
#>  -0.1167 -0.1300  0.8571  0.8235  0.8943  0.5631  0.5717  0.4546 -0.1864
#>   0.1433  0.2002 -0.4210  0.6734 -0.2746 -0.8153  0.6251  0.3999  0.6532
#> 
#> Columns 10 to 18 -0.2711  0.0145 -0.1932  0.7233 -0.0184  0.2043 -0.2979  0.1570 -0.1046
#>  -0.7779 -0.4321  0.0277  0.7265  0.4409  0.2673 -0.3034  0.2655  0.0309
#>  -0.8645  0.4395 -0.4064  0.4432 -0.3218 -0.5363 -0.1060  0.5155 -0.1965
#> 
#> Columns 19 to 20 -0.8511 -0.5662
#>  -0.6203  0.1272
#>  -0.0356  0.2744
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0029  0.2493  0.2794  0.1187 -0.0887 -0.1197  0.0403 -0.5404 -0.2387
#>   0.7215 -0.3128 -0.0018  0.5723  0.2840  0.1378  0.4114 -0.4605 -0.2429
#>   0.5773 -0.0674 -0.1265 -0.2128  0.3235 -0.0903 -0.3276 -0.0488 -0.5248
#> 
#> Columns 10 to 18 -0.1877  0.4155 -0.0611 -0.1181 -0.0857  0.0386  0.2441  0.3106  0.3313
#>   0.2373  0.2460 -0.0678 -0.1003 -0.6252  0.2790 -0.0884 -0.6123  0.5910
#>  -0.7056  0.0742  0.0942 -0.2264 -0.4008  0.4393 -0.3634  0.1297  0.7263
#> 
#> Columns 19 to 20 -0.3145 -0.0053
#>   0.4516 -0.6888
#>   0.1734  0.0751
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>