RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6811  0.0080  0.0143 -0.4335  0.4154  0.0574  0.5179  0.3476 -0.6631
#>   0.2147 -0.1566 -0.1882 -0.6809  0.1497 -0.0002 -0.0710  0.1028  0.2363
#>   0.0701 -0.3851  0.8263 -0.9518  0.4022 -0.7939 -0.6286 -0.1760  0.6673
#> 
#> Columns 10 to 18  0.3404 -0.0321  0.5344  0.5012  0.7726  0.6516  0.2608  0.6430  0.0565
#>   0.2331 -0.5759 -0.3384  0.5433  0.2316 -0.3651  0.0957  0.8053 -0.0658
#>   0.5961 -0.4552 -0.8937  0.6911 -0.2958 -0.5258 -0.3675 -0.3781 -0.0185
#> 
#> Columns 19 to 20 -0.4984  0.4413
#>  -0.4687 -0.4900
#>   0.4858  0.2506
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5808 -0.1351 -0.2420  0.5734 -0.3743  0.1207  0.7152 -0.5387  0.0119
#>  -0.5864 -0.0409 -0.1053 -0.7403  0.1365 -0.6007 -0.1753 -0.0051  0.0856
#>  -0.6096 -0.2216  0.4104  0.0683  0.1098  0.1732 -0.0134 -0.0504  0.1424
#> 
#> Columns 10 to 18  0.7254 -0.5839 -0.0231 -0.0265 -0.6051  0.4306 -0.5910  0.6697 -0.3053
#>  -0.4934 -0.0808 -0.4003 -0.0075 -0.2283  0.7324 -0.1186 -0.0977 -0.1635
#>  -0.3588 -0.6149 -0.3913  0.5837  0.0382  0.0403 -0.2992  0.5166  0.0642
#> 
#> Columns 19 to 20 -0.4746  0.0584
#>   0.1950  0.0049
#>  -0.1140  0.4054
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2754 -0.2489 -0.4275 -0.0388 -0.7170 -0.2490  0.1991 -0.4310  0.1705
#>  -0.5675  0.0412 -0.3294 -0.4132 -0.3416 -0.0902  0.0644 -0.5783  0.1237
#>  -0.4867 -0.1292  0.0973  0.4930 -0.1311 -0.1508  0.2328 -0.4987  0.2651
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4348  0.3896 -0.7306  0.5117  0.0665  0.1538  0.3409  0.6359 -0.1450
#>   0.6620 -0.1528  0.0359  0.1331 -0.0598 -0.0026 -0.7048  0.3571 -0.1132
#>   0.3805  0.2415  0.1411 -0.1143 -0.2585  0.6970 -0.6775 -0.7709  0.5887
#> 
#> Columns 10 to 18 -0.1812  0.1744 -0.4222  0.9279 -0.2895  0.3256  0.8052  0.2393 -0.1025
#>   0.5183  0.1709  0.1888  0.4754 -0.0440  0.4255 -0.2565 -0.7051 -0.1152
#>   0.3602  0.2418 -0.1337  0.0660  0.2152  0.1419  0.1020 -0.3892  0.5071
#> 
#> Columns 19 to 20  0.3162  0.7970
#>  -0.3498  0.3965
#>   0.1188  0.3586
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4122 -0.3151 -0.1418 -0.0028  0.1926 -0.4650 -0.1313 -0.1745  0.0874
#>  -0.4024 -0.2691 -0.0424 -0.1444 -0.2446  0.1999  0.3316 -0.3550 -0.0207
#>  -0.3369 -0.0345 -0.0957  0.2155 -0.0822 -0.1391  0.2398 -0.2414  0.2819
#> 
#> Columns 10 to 18 -0.0264  0.1497  0.0168 -0.1996 -0.0832  0.4604  0.1291  0.3153 -0.2000
#>   0.4613 -0.0201  0.0375 -0.2022 -0.1121  0.0169 -0.3773  0.5021 -0.2316
#>  -0.3905  0.1272 -0.4868 -0.2714 -0.3268  0.4009 -0.3831  0.3252 -0.1141
#> 
#> Columns 19 to 20 -0.1075  0.7306
#>  -0.2041  0.7641
#>   0.1244  0.3891
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>