RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4495 -0.3397 -0.1540  0.2666  0.7177  0.6723 -0.7592  0.0965 -0.2765
#>   0.2260 -0.8379  0.4079  0.3799 -0.3899 -0.0628  0.3455  0.7358  0.6931
#>   0.2608 -0.8791  0.7201  0.6015 -0.5460 -0.8647  0.7341 -0.6353 -0.5505
#> 
#> Columns 10 to 18 -0.8350 -0.5786  0.6180  0.2447 -0.3835 -0.5701 -0.5216 -0.4469 -0.5653
#>  -0.2438  0.8019  0.1272  0.3969  0.2915  0.1693 -0.4102  0.3485  0.2570
#>  -0.6662  0.4616  0.3559  0.0381 -0.4600 -0.6333  0.8482 -0.4013 -0.9304
#> 
#> Columns 19 to 20 -0.4472  0.2625
#>   0.1628  0.0482
#>   0.4230 -0.0834
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5124 -0.0834 -0.2236  0.6038 -0.2086 -0.3352  0.2127 -0.4051 -0.7149
#>   0.4046 -0.2826 -0.2810 -0.0393  0.4473 -0.1444  0.5879  0.2589 -0.0434
#>  -0.5222 -0.3028 -0.4543 -0.5812  0.0042  0.4694 -0.2175 -0.0196 -0.2005
#> 
#> Columns 10 to 18  0.1079 -0.1108 -0.4344  0.1126  0.2474  0.1063  0.4321 -0.0596  0.1840
#>   0.0491  0.5309  0.4172 -0.1287 -0.0976  0.6803 -0.1567  0.3163  0.4793
#>   0.4734 -0.2945 -0.0370  0.7222  0.5443  0.1905 -0.8008 -0.4033  0.7378
#> 
#> Columns 19 to 20  0.3464 -0.3746
#>   0.0579  0.1474
#>  -0.0597  0.2029
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3570 -0.3876 -0.2960  0.0797 -0.0946  0.5380 -0.3256 -0.1069  0.3272
#>   0.0541 -0.1004 -0.2864  0.2465 -0.4768 -0.1097  0.5059  0.2949  0.1295
#>  -0.0096 -0.3625  0.0852  0.4953 -0.1563 -0.1136 -0.1838  0.2183 -0.1376
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2581  0.8441  0.4544  0.7950 -0.4487  0.8718 -0.2395  0.0219  0.1488
#>   0.5773 -0.4660  0.7399 -0.0541  0.7532 -0.1236 -0.6882 -0.2043 -0.3064
#>  -0.1352 -0.4956  0.4105 -0.1072  0.6146 -0.0852  0.0161 -0.3416 -0.1936
#> 
#> Columns 10 to 18  0.5771 -0.6176  0.2341 -0.0869  0.7882  0.1188 -0.3893 -0.2431  0.0152
#>   0.6074 -0.1045 -0.8032  0.7451 -0.2749  0.6125 -0.1209 -0.2126  0.6281
#>   0.3722  0.0021  0.4281  0.3934  0.7276  0.2950 -0.2146  0.2294  0.6724
#> 
#> Columns 19 to 20  0.1519  0.5027
#>   0.7169 -0.1013
#>   0.0552 -0.3704
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1483 -0.6739 -0.7232  0.0851  0.1874  0.5644  0.5959 -0.4061 -0.4381
#>  -0.1139 -0.2170  0.3090 -0.0194 -0.0491 -0.2011  0.0301 -0.3285 -0.1762
#>  -0.2454 -0.4379 -0.1200  0.5775  0.1986  0.0509 -0.0011  0.1555 -0.5182
#> 
#> Columns 10 to 18  0.3614 -0.3725  0.3719  0.3910 -0.1291  0.5367  0.2750 -0.3559  0.0314
#>  -0.2914 -0.2349  0.0823  0.3183  0.3285  0.1949 -0.3137 -0.4158 -0.0759
#>  -0.2650 -0.2260  0.2417  0.3748  0.0411  0.3570  0.3064 -0.1352 -0.3865
#> 
#> Columns 19 to 20 -0.1102 -0.3570
#>   0.3412 -0.1577
#>   0.1699 -0.0913
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>