RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5371 -0.7035  0.6794  0.6783 -0.1898 -0.3128 -0.0239 -0.7250 -0.4349
#>   0.7330  0.3223  0.0303 -0.4882 -0.4577  0.1644  0.1145 -0.7306 -0.8231
#>   0.3669  0.7875  0.2603  0.9096 -0.5869  0.1142  0.3975  0.4411 -0.5291
#> 
#> Columns 10 to 18  0.7345  0.5584  0.9075  0.1464 -0.3499 -0.2711 -0.7922  0.9214 -0.4729
#>   0.7749  0.5062 -0.4086 -0.7202 -0.5195  0.0970  0.6866  0.0441  0.8903
#>   0.1605  0.5578 -0.1139  0.7546 -0.7411 -0.8277  0.0730  0.3590 -0.9365
#> 
#> Columns 19 to 20 -0.1548 -0.6317
#>   0.1212 -0.4490
#>   0.0387 -0.5102
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3571 -0.2573 -0.0492 -0.1093  0.0783 -0.5384 -0.1136 -0.6845 -0.5016
#>  -0.0552  0.0660 -0.1655 -0.0503 -0.5385  0.4281  0.0375 -0.5935 -0.8564
#>   0.0889 -0.4298 -0.7751 -0.2621 -0.0169 -0.4287  0.2608 -0.7696  0.4040
#> 
#> Columns 10 to 18  0.0867  0.2928  0.5588  0.2725 -0.3250 -0.3979 -0.1237  0.0224  0.3811
#>   0.1399 -0.0531 -0.3646  0.1935 -0.5666 -0.1011 -0.2590  0.0589  0.4005
#>  -0.4392 -0.2530  0.4731  0.2672 -0.7023 -0.2468 -0.3889 -0.1396  0.3780
#> 
#> Columns 19 to 20  0.1402 -0.0953
#>   0.4934 -0.6800
#>  -0.4443  0.1476
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0868  0.1438 -0.2454 -0.1468 -0.4471 -0.5495 -0.0322 -0.6331 -0.4437
#>  -0.1170  0.0174 -0.1595 -0.0364 -0.4493 -0.2083 -0.0403 -0.5304 -0.5004
#>   0.2678  0.0392  0.0761 -0.1059 -0.3759 -0.1087 -0.0746 -0.4563 -0.6538
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2587 -0.1462  0.1234 -0.0830 -0.2280  0.0164  0.4294 -0.4055  0.1864
#>   0.1860  0.3314  0.1907  0.3880 -0.3052  0.0124  0.0096  0.1082 -0.4077
#>  -0.0031  0.4311  0.6934 -0.0505 -0.5017  0.5807 -0.2993 -0.3960  0.3292
#> 
#> Columns 10 to 18 -0.5983 -0.6888 -0.1699  0.1342  0.2509 -0.0395  0.3868  0.2984  0.0155
#>   0.0748 -0.2113 -0.0726 -0.2704  0.2957 -0.0250 -0.2838 -0.4517  0.0179
#>   0.1014 -0.4338  0.1492  0.6007  0.2548 -0.5585  0.1884 -0.1046 -0.7206
#> 
#> Columns 19 to 20  0.2637 -0.1062
#>  -0.0311  0.1206
#>  -0.6125 -0.3384
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2716  0.1728 -0.2679  0.0485 -0.3038  0.0690 -0.1393 -0.6050 -0.5517
#>   0.3019  0.1473 -0.0881 -0.0368 -0.0450 -0.3994  0.3269 -0.3667 -0.2998
#>   0.2337 -0.0439 -0.4031 -0.3992 -0.0306  0.1231  0.0783 -0.5563 -0.4630
#> 
#> Columns 10 to 18 -0.0716  0.3050  0.0472  0.1264 -0.4901 -0.3327  0.0564 -0.3602  0.4999
#>  -0.2783 -0.1695 -0.0008  0.0264 -0.3876 -0.1887 -0.1276 -0.1395 -0.1818
#>  -0.2252  0.2889 -0.2129  0.0114 -0.5628 -0.2018  0.1172 -0.0803  0.3384
#> 
#> Columns 19 to 20 -0.0343 -0.3365
#>  -0.1102  0.3237
#>  -0.0244 -0.4473
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>