RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3845  0.5662  0.7796 -0.2014  0.0781  0.9362  0.0215 -0.5812  0.4508
#>   0.5489  0.0545 -0.0511 -0.7683  0.2314  0.4700 -0.4848  0.0488  0.8574
#>   0.0431 -0.9114 -0.1695 -0.6605 -0.8183 -0.4680 -0.8877  0.8181 -0.5947
#> 
#> Columns 10 to 18 -0.1197 -0.4574  0.2872 -0.5911 -0.5693  0.3347  0.4743 -0.8324  0.8175
#>  -0.3567 -0.3474 -0.1416  0.6735  0.2279 -0.1347 -0.9078  0.9380 -0.8790
#>   0.7042  0.5631 -0.1231 -0.4630  0.4834  0.9347  0.2071  0.0123 -0.3763
#> 
#> Columns 19 to 20 -0.4154  0.5631
#>  -0.0094 -0.5561
#>  -0.0466 -0.0716
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1330  0.2212  0.0325 -0.1509 -0.2679  0.4343 -0.3484 -0.0925 -0.1551
#>  -0.0111  0.2847  0.0880 -0.1423 -0.2088  0.4291  0.2189 -0.4788  0.0903
#>  -0.2895 -0.0274  0.3352  0.0921  0.4759 -0.2833  0.2387 -0.1753  0.0323
#> 
#> Columns 10 to 18 -0.1389 -0.4403 -0.6177  0.1224  0.4323  0.4748 -0.0911  0.1080  0.4169
#>   0.1111 -0.4120  0.7693 -0.5457 -0.6650  0.4554 -0.1505 -0.4767  0.3071
#>   0.0251  0.3789  0.6709 -0.8436 -0.1636 -0.2940 -0.0711  0.1266  0.3652
#> 
#> Columns 19 to 20 -0.7806  0.0431
#>   0.3434 -0.6357
#>  -0.4789  0.2969
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0321 -0.1373  0.0267 -0.0983  0.2349  0.4072 -0.3520  0.1277 -0.4677
#>   0.4248  0.3840 -0.1205 -0.6041  0.4559  0.4602  0.0615  0.0247 -0.0661
#>  -0.0933 -0.2812  0.3764 -0.3288  0.1631  0.2576 -0.5554  0.3267  0.1410
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0241 -0.4496 -0.2552  0.2180  0.5039 -0.3265  0.3773  0.0412  0.6142
#>  -0.1651 -0.1864  0.1403  0.0392  0.5375 -0.3358  0.3934  0.3775 -0.1623
#>   0.0566  0.0041  0.0746 -0.0644 -0.1995 -0.5963 -0.1303  0.1878  0.0294
#> 
#> Columns 10 to 18  0.5610  0.1901 -0.4915 -0.0948  0.1281  0.0151  0.3430  0.2423 -0.2520
#>  -0.6778  0.2006  0.2836 -0.5991  0.3355 -0.2138  0.1004 -0.3616  0.1183
#>  -0.2538 -0.5042 -0.1358 -0.6726  0.1039 -0.5237  0.2954  0.1173  0.5165
#> 
#> Columns 19 to 20 -0.2521 -0.6506
#>  -0.5995 -0.3476
#>  -0.1306 -0.2343
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0672 -0.2027  0.1972 -0.5741  0.0450  0.2441 -0.1858  0.2814 -0.4838
#>   0.3276 -0.2033  0.2349 -0.3620  0.1842  0.5093 -0.2919  0.4304  0.1670
#>  -0.0868 -0.3783  0.0565 -0.4074  0.2184  0.1199 -0.0338  0.2808  0.1788
#> 
#> Columns 10 to 18  0.2680 -0.2032  0.2475 -0.4092 -0.3334  0.5936 -0.5463  0.2296  0.0036
#>  -0.2609  0.3919 -0.0031 -0.3513  0.2137  0.5407 -0.2957 -0.0409  0.1409
#>  -0.3072 -0.0251  0.2498 -0.3849 -0.0282  0.5161 -0.3917  0.3158  0.1199
#> 
#> Columns 19 to 20 -0.7627 -0.3675
#>  -0.4385  0.1796
#>  -0.5909 -0.3136
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>