RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1364  0.2137 -0.2140 -0.1728  0.4766  0.6356  0.2895  0.2558 -0.2743
#>  -0.2069  0.3703  0.5695  0.0581  0.0626  0.4926  0.5903  0.2407 -0.0775
#>  -0.2493 -0.5304  0.0254 -0.1021  0.5831  0.9402 -0.3481  0.1873 -0.6038
#> 
#> Columns 10 to 18  0.2026 -0.2206  0.6505 -0.8284  0.0939  0.6450 -0.3894  0.2517  0.3347
#>   0.4235  0.0464  0.8384 -0.1404 -0.2556  0.9612 -0.0219  0.0923  0.3519
#>  -0.3309  0.1608  0.7030 -0.6175  0.2143  0.9495  0.5169 -0.4197 -0.3343
#> 
#> Columns 19 to 20  0.0083  0.3474
#>  -0.8448  0.1529
#>  -0.6281  0.4154
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4150  0.3939  0.1476  0.1776  0.0978  0.4845 -0.3520  0.2891 -0.3521
#>  -0.0158  0.2765  0.2861  0.0808  0.3540  0.1070  0.0205  0.4789  0.1455
#>  -0.4133  0.1856  0.5397 -0.1439  0.5626  0.2635  0.3502  0.5408 -0.0819
#> 
#> Columns 10 to 18 -0.1459 -0.0416  0.4392  0.2315  0.3775  0.7722  0.1646 -0.1079  0.1655
#>   0.4375 -0.3221  0.2383  0.2243  0.3457  0.3668 -0.0785 -0.1180  0.1333
#>   0.2586 -0.4577  0.6712  0.1122  0.3287  0.5256  0.3253 -0.3032  0.7507
#> 
#> Columns 19 to 20 -0.7151 -0.0145
#>  -0.5952 -0.0230
#>  -0.5121  0.0069
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.6300  0.1567  0.4627  0.2633  0.4635  0.1382  0.2850  0.7640 -0.1272
#>  -0.2665 -0.2624  0.1342  0.1070  0.2955 -0.1590  0.5355  0.4990 -0.0497
#>  -0.2726  0.3069 -0.0528  0.1433  0.0603  0.4548  0.3198  0.3459  0.3988
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3125  0.4257 -0.3380  0.0645  0.2787 -0.9098 -0.7560  0.4303 -0.0452
#>  -0.5978 -0.2009  0.2753 -0.3301  0.6571 -0.9664 -0.2754  0.4289 -0.1933
#>   0.6003  0.1407  0.2274 -0.5264  0.4940  0.3620  0.7045 -0.7387 -0.4867
#> 
#> Columns 10 to 18  0.0501 -0.5435 -0.4765 -0.1362  0.6726  0.0767  0.0305  0.7365  0.7327
#>  -0.7872 -0.6561 -0.5369 -0.3267  0.3580 -0.3019  0.0976  0.2372  0.7686
#>   0.5938  0.1831  0.3985  0.0653 -0.6788  0.1499 -0.3525 -0.1055 -0.6862
#> 
#> Columns 19 to 20  0.4243 -0.4104
#>  -0.1263 -0.6163
#>   0.1727  0.1796
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3824 -0.3813  0.2420  0.0032  0.5488 -0.2786  0.4194  0.2160 -0.0144
#>   0.0137 -0.2264  0.2138  0.4296  0.1171  0.2211  0.6309  0.0646 -0.2466
#>  -0.2264  0.0139  0.4336  0.1751 -0.0246 -0.2848  0.2681  0.4623 -0.2494
#> 
#> Columns 10 to 18  0.4632 -0.0720 -0.0166  0.0434  0.1437  0.2421  0.3604  0.2171  0.4024
#>   0.6166 -0.1691  0.5003  0.2480  0.2971  0.0961 -0.0152  0.2821 -0.1392
#>   0.0771 -0.1565  0.4703  0.1641  0.6112  0.6831  0.1223  0.0842  0.3240
#> 
#> Columns 19 to 20 -0.4982  0.5534
#>  -0.7295  0.5302
#>  -0.2572 -0.2746
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>