RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1711 -0.3317  0.0653 -0.1612  0.1426  0.4884 -0.0641  0.1778 -0.2107
#>   0.9027 -0.9740  0.0396  0.2956  0.8095 -0.2469  0.6736  0.0504 -0.3217
#>  -0.7500  0.0370 -0.6160  0.1203 -0.3643  0.3336  0.0364  0.0109 -0.2834
#> 
#> Columns 10 to 18  0.6257  0.5444 -0.6765 -0.4008 -0.2572 -0.0034 -0.0540 -0.3281  0.8802
#>   0.2386 -0.0689  0.1380 -0.2090  0.2804 -0.7814 -0.1531  0.9028 -0.3470
#>  -0.6606  0.9200 -0.4547  0.9569  0.7862 -0.3529 -0.4483  0.6055 -0.8256
#> 
#> Columns 19 to 20  0.7429 -0.7578
#>   0.3195 -0.0751
#>   0.4604 -0.4414
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2672 -0.5599 -0.2005  0.1659 -0.2348  0.1813 -0.1952 -0.4438 -0.4287
#>   0.3882 -0.4744 -0.3137 -0.3319 -0.4164  0.4604  0.1543 -0.6483 -0.1832
#>   0.6431  0.3778 -0.1276  0.0340 -0.6007 -0.1451 -0.4538  0.0973 -0.6420
#> 
#> Columns 10 to 18  0.5653  0.1367 -0.0965  0.4114  0.3502 -0.5984 -0.8004 -0.1529 -0.3150
#>   0.3932  0.1836  0.4405  0.5179  0.3177 -0.8138 -0.5866  0.5841  0.3080
#>  -0.1642  0.3060  0.3276 -0.5727 -0.5207  0.0395 -0.6037 -0.5739  0.1371
#> 
#> Columns 19 to 20 -0.0438 -0.1365
#>  -0.2801 -0.2224
#>  -0.0795 -0.4161
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1647 -0.2027 -0.4194 -0.1683 -0.6928  0.2903 -0.4078 -0.3956 -0.4741
#>  -0.1958 -0.0944 -0.2636 -0.0603 -0.6797  0.2420 -0.0766 -0.5049 -0.1485
#>   0.4356 -0.1294  0.1515  0.2077 -0.4801  0.5743  0.1878 -0.2127 -0.3209
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4971 -0.2593  0.1795 -0.6336  0.4081 -0.0407 -0.1393 -0.1929 -0.2465
#>   0.3828  0.2628 -0.6498  0.3461 -0.2172  0.6650 -0.1116  0.7002  0.2388
#>   0.1838  0.0817  0.1658 -0.0149  0.0174 -0.1980 -0.2209 -0.4225 -0.6202
#> 
#> Columns 10 to 18  0.4069 -0.1157 -0.2074  0.1363 -0.2309  0.1856 -0.0178  0.0798 -0.6013
#>   0.1492  0.1623 -0.4842 -0.1582 -0.3017 -0.1718 -0.5946 -0.4976  0.1673
#>   0.0475 -0.2116 -0.3164 -0.3561 -0.1022 -0.4765  0.7806  0.6892 -0.5665
#> 
#> Columns 19 to 20  0.3014 -0.2900
#>  -0.1891 -0.1756
#>  -0.1049  0.1207
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2458 -0.2216 -0.3340  0.0747 -0.4862  0.2195 -0.5486 -0.0346 -0.6963
#>  -0.0595  0.0384  0.0179  0.0901 -0.2641  0.3326 -0.1246 -0.3393  0.0070
#>  -0.0338 -0.3813 -0.4271  0.0448 -0.6046  0.3217 -0.4175 -0.3592 -0.5700
#> 
#> Columns 10 to 18  0.2609  0.0288  0.2413  0.1693 -0.0377 -0.5091 -0.4823 -0.1261  0.1833
#>   0.4073  0.0307  0.4227  0.2303 -0.0329  0.0469 -0.2198 -0.3159  0.0326
#>   0.1224  0.3590 -0.0392  0.1649  0.0873 -0.6785 -0.5685  0.3766  0.1262
#> 
#> Columns 19 to 20 -0.0315 -0.2035
#>  -0.0933  0.0475
#>   0.0722 -0.3335
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>