RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5168  0.6070  0.1052 -0.0256 -0.3927 -0.4395  0.0183 -0.1736 -0.4536
#>   0.5682  0.3171 -0.1928  0.2712 -0.0351  0.4700  0.0079 -0.1198  0.0969
#>   0.3621 -0.5936  0.2602  0.1677 -0.0838  0.8200  0.2941 -0.3100 -0.2989
#> 
#> Columns 10 to 18 -0.4852 -0.1015 -0.1149 -0.6681  0.5614  0.4312  0.6402 -0.1284  0.3914
#>   0.2728  0.2482 -0.2820 -0.1119 -0.0690 -0.6237  0.1076  0.7508 -0.1383
#>  -0.2469  0.7065  0.3545  0.2754  0.2790 -0.8577 -0.5187  0.2246  0.7184
#> 
#> Columns 19 to 20  0.9093 -0.7881
#>   0.7802 -0.7081
#>  -0.1855 -0.1665
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1075  0.3097 -0.0130 -0.4816  0.5575  0.4893 -0.1252 -0.3976 -0.2420
#>  -0.0728 -0.0275 -0.1106  0.2364  0.0155  0.1685  0.3850  0.2057  0.5455
#>   0.1918 -0.1819  0.0158  0.2221  0.3172  0.5682  0.3134 -0.1875 -0.0542
#> 
#> Columns 10 to 18 -0.0751  0.4673 -0.1861 -0.4140 -0.2039 -0.0498 -0.5985  0.1811  0.1462
#>  -0.0133  0.1590 -0.0458  0.0179 -0.3113  0.5699  0.0106  0.0078  0.2471
#>   0.2760  0.2230  0.2708  0.5692 -0.2148 -0.2442  0.2754  0.1274  0.5905
#> 
#> Columns 19 to 20  0.5614 -0.2109
#>  -0.0759 -0.2337
#>   0.8348 -0.7057
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2744 -0.1821 -0.2304 -0.4170  0.1006  0.1495  0.7596 -0.1577  0.0642
#>   0.2419 -0.0751 -0.3966 -0.0571  0.1900  0.6261  0.3539  0.0270  0.0213
#>   0.2337 -0.2793 -0.3675  0.1017  0.1662  0.4481  0.3269 -0.3383  0.2901
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4181 -0.4424  0.1601  0.3808  0.4501 -0.3435 -0.4294  0.6717 -0.3861
#>   0.6811  0.0142 -0.1601  0.2760  0.7195 -0.6634 -0.5120 -0.5663 -0.2249
#>  -0.4585 -0.6322 -0.5900 -0.0630  0.8262 -0.1583 -0.7321 -0.2319 -0.6071
#> 
#> Columns 10 to 18 -0.0716 -0.0694  0.3126  0.1048 -0.4045 -0.1407  0.2757  0.0455 -0.2313
#>   0.2541  0.2850 -0.2677  0.5227 -0.5980 -0.4474 -0.8130 -0.7196 -0.8279
#>  -0.3380 -0.4792  0.1737 -0.2104 -0.5738 -0.5646 -0.7912  0.2834  0.2215
#> 
#> Columns 19 to 20  0.3035  0.1407
#>   0.5027  0.1667
#>   0.3826  0.1064
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0704  0.1575 -0.6554 -0.1424  0.0702  0.3182  0.3173  0.4857  0.5430
#>  -0.0777  0.0187 -0.6553 -0.1414  0.4449  0.5616  0.3528  0.4007  0.7037
#>   0.1406 -0.2819 -0.5204  0.1824 -0.0325  0.5144  0.4459  0.1963  0.6889
#> 
#> Columns 10 to 18 -0.1116  0.3703  0.0649 -0.2201  0.1516  0.4642 -0.2043  0.0475  0.0175
#>   0.0607  0.4912 -0.2387  0.3785  0.3671  0.3108 -0.0654 -0.2135  0.3372
#>   0.0799  0.4301 -0.1181  0.3632  0.1737  0.3930 -0.0767  0.0403  0.1603
#> 
#> Columns 19 to 20  0.0928  0.0118
#>   0.2769 -0.0739
#>  -0.2414 -0.4148
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>