RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7369  0.6902  0.4225  0.7701 -0.5707 -0.4686 -0.4772  0.5574  0.3702
#>   0.5630  0.8027 -0.4510  0.6027 -0.3364  0.8788 -0.8219  0.3663 -0.1581
#>   0.2765  0.2475 -0.7021  0.2867 -0.5938  0.0370 -0.3787 -0.4366 -0.4965
#> 
#> Columns 10 to 18  0.6025 -0.8726 -0.4028  0.5393 -0.0502  0.1010 -0.5388 -0.4326  0.8652
#>   0.8249 -0.4776  0.0902  0.1096 -0.5276  0.5757 -0.8384 -0.6203 -0.4631
#>  -0.1898 -0.7225 -0.3412 -0.2860 -0.5936 -0.3398  0.6917 -0.6856 -0.1173
#> 
#> Columns 19 to 20  0.0266 -0.6734
#>   0.1636  0.2091
#>  -0.6307  0.4612
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1220 -0.0798 -0.3401  0.2660 -0.4190 -0.1022 -0.2746  0.0137 -0.0500
#>   0.0074 -0.0602  0.3421  0.3688 -0.6164  0.3578 -0.0117  0.5466  0.3278
#>   0.1005  0.5367 -0.0682 -0.6415  0.3518 -0.4161  0.5510 -0.3750 -0.1188
#> 
#> Columns 10 to 18 -0.0745 -0.7464 -0.1205 -0.0933 -0.2051 -0.4369  0.6574 -0.3745  0.3730
#>  -0.0251 -0.0935  0.0043  0.1463 -0.5506 -0.1609  0.1004 -0.6760  0.4342
#>   0.1350 -0.0103 -0.1328 -0.1377 -0.7909 -0.5999  0.3582  0.4172  0.0381
#> 
#> Columns 19 to 20 -0.2382  0.1066
#>   0.5984 -0.3007
#>  -0.0562  0.2868
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3410 -0.1519 -0.5360 -0.3202  0.0792 -0.0887  0.1532 -0.0842  0.0067
#>   0.0538 -0.0493 -0.4795 -0.2707 -0.2018 -0.5019  0.3576  0.2607 -0.4066
#>   0.1755  0.6084 -0.4637 -0.5326 -0.1840 -0.4582  0.3990 -0.1323  0.0323
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1010  0.1981  0.2722  0.7442 -0.6200 -0.7780  0.2305 -0.6500  0.1717
#>  -0.0244 -0.1879 -0.4309  0.2144  0.0252  0.6247  0.3379 -0.3101  0.3907
#>  -0.7830  0.6632 -0.5290  0.6169  0.0692 -0.5061  0.0828 -0.0427  0.0193
#> 
#> Columns 10 to 18  0.7666 -0.1061 -0.3105  0.4346 -0.6862  0.4721  0.0566 -0.0727  0.1882
#>  -0.1008  0.7690 -0.7044 -0.2119 -0.2409 -0.0541 -0.3912  0.8278  0.4951
#>   0.3818  0.6948 -0.1706  0.2525 -0.6854  0.1401 -0.1406  0.5569  0.1553
#> 
#> Columns 19 to 20  0.1677  0.5243
#>  -0.2946 -0.1399
#>  -0.3445  0.1595
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1336  0.5583 -0.0644 -0.3004  0.2356 -0.3767  0.3684  0.2549 -0.4381
#>   0.4228  0.5668 -0.4153 -0.4544 -0.4596 -0.3050  0.1236 -0.1034 -0.1429
#>   0.3670  0.3756 -0.1941 -0.3841  0.1626 -0.1155 -0.1901 -0.0051 -0.0175
#> 
#> Columns 10 to 18  0.5190 -0.0204  0.2044 -0.2487 -0.6789 -0.0483  0.1233 -0.1092 -0.1381
#>   0.4763 -0.4987 -0.2780  0.1644 -0.4801 -0.4733  0.4247  0.1651  0.0275
#>   0.2103 -0.1190  0.1278 -0.2445 -0.4782 -0.5079  0.4309 -0.2202  0.1794
#> 
#> Columns 19 to 20  0.0632 -0.1818
#>  -0.0455 -0.0925
#>  -0.0643  0.1379
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>