RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5288 -0.5133 -0.0588  0.6859  0.6552  0.4015 -0.5190  0.6700 -0.6734
#>  -0.3890 -0.5422  0.8343 -0.2204  0.9122  0.5523 -0.3016 -0.1138  0.3760
#>  -0.2615  0.1018 -0.2807  0.5783  0.6774  0.6647 -0.3351 -0.5544 -0.5520
#> 
#> Columns 10 to 18 -0.4103  0.0978  0.9540  0.4142  0.5198 -0.1543  0.7641  0.7700  0.1040
#>   0.5781  0.4323 -0.3654  0.6206  0.0160 -0.3410 -0.2721 -0.4083  0.2800
#>   0.3245  0.6603  0.2175  0.1895  0.4472  0.7729  0.4526 -0.0951 -0.5446
#> 
#> Columns 19 to 20 -0.6022  0.6488
#>  -0.1545  0.1631
#>  -0.5987  0.1054
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0207 -0.4867 -0.1163  0.3282  0.3749  0.7454 -0.5512  0.0281 -0.4357
#>   0.0592  0.4786  0.1293  0.7318  0.5967  0.3037  0.4320  0.3960 -0.4475
#>  -0.1518 -0.0368  0.2861 -0.0522  0.6154  0.2062  0.1220 -0.2419  0.1697
#> 
#> Columns 10 to 18 -0.1756 -0.0455  0.0044  0.1097  0.0707  0.0412  0.3643  0.2535 -0.5904
#>   0.6010 -0.1266  0.1070  0.6558  0.2779 -0.1556  0.3844  0.4717 -0.5106
#>   0.5548  0.3109  0.0538 -0.1771  0.2866 -0.2155  0.2017  0.4113  0.0953
#> 
#> Columns 19 to 20 -0.2480  0.0296
#>  -0.1110  0.3659
#>  -0.3007 -0.5061
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.7164 -0.4312  0.1426  0.2108  0.6965  0.5255 -0.1904 -0.4153  0.3145
#>   0.0438 -0.1360 -0.5034  0.3096  0.3851  0.7699  0.1604  0.5970  0.0187
#>  -0.0102  0.3034 -0.1834  0.5246  0.5070  0.5534  0.3609  0.3118  0.1165
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4178  0.6779  0.3609 -0.5670  0.2984 -0.0113  0.6129  0.2644 -0.3662
#>  -0.8827  0.3608  0.1104 -0.3603  0.3925  0.6573  0.5738 -0.2088 -0.6116
#>  -0.7909  0.7530  0.9295 -0.0702 -0.0768 -0.1016  0.5673  0.0736 -0.3873
#> 
#> Columns 10 to 18  0.2038  0.0935  0.1811  0.0499  0.4634 -0.4110 -0.5662 -0.3577  0.1473
#>   0.0777 -0.6369 -0.4293  0.1868 -0.2533 -0.6017 -0.1344 -0.1029  0.3660
#>   0.5568  0.5322  0.4108  0.1848  0.2132 -0.6077 -0.6211  0.6581  0.5252
#> 
#> Columns 19 to 20  0.3967  0.3666
#>  -0.4239 -0.2060
#>   0.0700 -0.1301
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1935  0.4499 -0.0630  0.3435  0.3381  0.3166  0.1481  0.6517 -0.1912
#>  -0.3173 -0.1522 -0.0332  0.1094  0.7270  0.5785 -0.2419  0.3271  0.3084
#>  -0.3957  0.0220 -0.2992  0.0507  0.2650  0.5988 -0.3027  0.7517  0.0823
#> 
#> Columns 10 to 18 -0.1044 -0.1266 -0.2640  0.3106 -0.3103  0.0046  0.5789  0.5557 -0.4541
#>   0.0545  0.0935  0.0145  0.2016 -0.3801 -0.4741 -0.5998  0.4198  0.0798
#>  -0.3787 -0.4047 -0.3512  0.3994 -0.5657 -0.6830  0.2317  0.2660 -0.4599
#> 
#> Columns 19 to 20 -0.3052  0.1950
#>  -0.6267  0.3661
#>  -0.3962  0.2240
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>