RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5482 -0.3557  0.6690  0.8161  0.7455  0.2076  0.0390  0.5230 -0.3420
#>   0.5577  0.9362 -0.0094 -0.5208 -0.2772  0.0178  0.4705 -0.2432 -0.8395
#>   0.7695  0.4351 -0.0260  0.8022  0.2138  0.2405 -0.2412  0.8043 -0.8693
#> 
#> Columns 10 to 18  0.6292  0.4312 -0.7485 -0.3079  0.6522 -0.4323 -0.6656 -0.4129  0.1528
#>   0.9375  0.1605 -0.3135 -0.8295 -0.0696  0.1096 -0.4786 -0.5682  0.6012
#>   0.3281  0.0306 -0.7752 -0.7853  0.8791 -0.3549  0.1950 -0.3640 -0.2994
#> 
#> Columns 19 to 20 -0.6767  0.6995
#>  -0.1751  0.5388
#>  -0.3373  0.4844
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4419  0.4888 -0.5922 -0.2342  0.1465  0.5172  0.1503  0.3459 -0.1540
#>   0.4450  0.3241 -0.2671  0.2374  0.0241 -0.1256 -0.5049 -0.2118  0.3283
#>   0.6245 -0.2622 -0.3658  0.0835  0.0708  0.1271 -0.1095  0.0783 -0.2271
#> 
#> Columns 10 to 18  0.6010  0.1352  0.1560 -0.4936 -0.2512  0.6239 -0.6611 -0.1546  0.5710
#>   0.4036 -0.1957  0.6324 -0.6431 -0.4666  0.4314 -0.1478  0.0531  0.6780
#>   0.5223 -0.3847  0.2360 -0.4002 -0.0587  0.6907 -0.3405  0.0585  0.0840
#> 
#> Columns 19 to 20 -0.2948  0.1788
#>  -0.3969  0.3883
#>   0.0130  0.4136
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.6457  0.2694  0.3223  0.1082 -0.0244 -0.1539  0.3228 -0.2418  0.1064
#>   0.7626 -0.1278  0.3096  0.3526 -0.2530 -0.0076  0.2404 -0.2047 -0.4850
#>   0.6253  0.2778 -0.1280 -0.1800 -0.2407 -0.1078  0.3752 -0.0226  0.3295
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3302 -0.3942 -0.3768  0.7078 -0.1366 -0.6380 -0.0354 -0.4059  0.6505
#>  -0.1376  0.4978  0.1537  0.0199 -0.0803 -0.8085  0.6950  0.3657 -0.1544
#>   0.2408  0.8028 -0.3289  0.0290 -0.2351 -0.6460  0.5285 -0.1009  0.0167
#> 
#> Columns 10 to 18  0.0113  0.7405  0.1561  0.1401 -0.1978 -0.2514 -0.0404  0.2111  0.7292
#>  -0.3739  0.4523  0.3612 -0.2794 -0.2555 -0.2890  0.5536  0.5178  0.6236
#>   0.0143  0.1025  0.5937 -0.3976 -0.3361  0.5071 -0.1769 -0.1276 -0.3995
#> 
#> Columns 19 to 20 -0.0598  0.5466
#>  -0.1751 -0.5087
#>   0.0941 -0.0107
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.8271 -0.1887 -0.5236  0.4102  0.1560  0.3102  0.3417  0.4172 -0.1106
#>   0.8579  0.0116  0.4208  0.3589 -0.0136  0.1318  0.4066  0.2449  0.0048
#>   0.6572  0.3569 -0.3535  0.3501  0.2043  0.4428  0.1767  0.0857  0.0802
#> 
#> Columns 10 to 18 -0.2856  0.3724 -0.0002 -0.5986 -0.1254 -0.1653 -0.1609 -0.5676  0.3899
#>   0.0254  0.1307  0.1081 -0.4821  0.4517 -0.2631 -0.3751 -0.3104  0.3021
#>   0.5849  0.1295  0.3123 -0.5363  0.1815  0.1788 -0.7081  0.0087  0.5704
#> 
#> Columns 19 to 20 -0.1399  0.3804
#>  -0.2811  0.6016
#>  -0.3778  0.4439
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>