RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6985  0.9515  0.5359  0.4878 -0.3638  0.4106 -0.7773 -0.7545  0.0240
#>  -0.9226 -0.6382  0.1298 -0.8600  0.8006  0.2281  0.5700 -0.3159 -0.8955
#>   0.0260  0.7598 -0.2921 -0.7488 -0.3388  0.4772  0.6422 -0.9693 -0.2092
#> 
#> Columns 10 to 18 -0.4059 -0.8046 -0.7076 -0.3026  0.9378  0.3472  0.8012  0.9542  0.8055
#>  -0.3371 -0.6680 -0.2114  0.0198 -0.4732  0.3590 -0.7961  0.4717  0.3474
#>  -0.4953  0.2362  0.4830  0.7448  0.7581  0.2275  0.8907  0.9446  0.0733
#> 
#> Columns 19 to 20 -0.9690 -0.2874
#>   0.9655 -0.4106
#>  -0.5109  0.0596
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6242  0.3798 -0.0561 -0.8875  0.3430 -0.3804  0.2652 -0.4918 -0.3246
#>  -0.3001 -0.3986  0.1910 -0.3509 -0.3990 -0.4424  0.4027 -0.3443  0.0213
#>   0.0180  0.6498  0.1250 -0.6023  0.3639 -0.6380  0.3939 -0.7941 -0.3521
#> 
#> Columns 10 to 18 -0.6206  0.1414 -0.1503  0.2448  0.1676  0.3247  0.4834  0.4435  0.0787
#>   0.4363 -0.6512  0.3414  0.1017 -0.0718 -0.5088 -0.1545  0.8059 -0.0527
#>  -0.5344 -0.2811  0.1615  0.3590 -0.2858 -0.0896  0.1506  0.4378 -0.0007
#> 
#> Columns 19 to 20 -0.6571  0.0128
#>   0.6772 -0.4913
#>  -0.8086  0.2362
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3982  0.0662 -0.1010 -0.4821  0.0303 -0.4442  0.1428 -0.3140 -0.4535
#>  -0.6877 -0.4498 -0.2021 -0.2900  0.3572 -0.1377  0.2124 -0.4617 -0.5097
#>  -0.4887 -0.1284 -0.0387 -0.5012 -0.2602 -0.7184  0.1384 -0.2061 -0.2668
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0429  0.4052  0.1016 -0.5246  0.6794  0.2705 -0.4741 -0.0217  0.0838
#>  -0.3270  0.1862  0.0886  0.5839  0.6109 -0.3917 -0.2038 -0.4430  0.1673
#>  -0.6059  0.6968  0.7792 -0.5013  0.2507 -0.2517  0.1157 -0.4201  0.2948
#> 
#> Columns 10 to 18 -0.3930 -0.2167  0.4765  0.3057  0.6478 -0.0163  0.3251 -0.3625 -0.5044
#>  -0.0234 -0.6006  0.5226 -0.2582  0.4332 -0.7437  0.2525  0.4857 -0.1224
#>  -0.5137 -0.5914  0.5598  0.1826  0.4499 -0.2928  0.7525  0.3359 -0.1003
#> 
#> Columns 19 to 20 -0.0877 -0.4791
#>   0.7120 -0.6507
#>   0.4417  0.6237
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2648 -0.3620  0.1168 -0.1621 -0.0861 -0.0666  0.2599 -0.2679 -0.3090
#>  -0.5505 -0.4959  0.2583 -0.2816 -0.0911 -0.5715  0.1234 -0.0718 -0.4041
#>  -0.4998 -0.3843  0.1553 -0.2473 -0.1989 -0.3040  0.5193 -0.8302 -0.2718
#> 
#> Columns 10 to 18  0.2588 -0.3186  0.2124  0.1648 -0.1714 -0.3452  0.0837  0.6643  0.3072
#>   0.1353 -0.2036  0.1097  0.5117 -0.5732 -0.6978 -0.0260  0.4133 -0.0652
#>   0.2805  0.0925  0.2170  0.2852  0.0025 -0.6608 -0.1703  0.7285 -0.1927
#> 
#> Columns 19 to 20  0.6511  0.0220
#>   0.6721 -0.4214
#>   0.1065  0.0012
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>