RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3002 -0.8566 -0.4047 -0.1330  0.2942  0.2024 -0.1087 -0.7107  0.3024
#>   0.3280 -0.8543 -0.0319 -0.7170 -0.9020  0.5183 -0.0471 -0.5734  0.2833
#>  -0.2670 -0.6634 -0.8122 -0.9270 -0.5847 -0.4945  0.9024  0.3129  0.7371
#> 
#> Columns 10 to 18  0.4179  0.3682  0.5491  0.1031  0.7557  0.5838 -0.5270 -0.9096 -0.0811
#>  -0.6015 -0.5359  0.1104  0.2355 -0.0580  0.1173 -0.5211 -0.4922  0.0020
#>  -0.6651 -0.5635  0.7454 -0.2051  0.2375 -0.6847 -0.2969 -0.8717 -0.6962
#> 
#> Columns 19 to 20 -0.7740 -0.8919
#>  -0.2584  0.5462
#>  -0.5297 -0.5189
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1685 -0.0148  0.2390  0.1834  0.1414 -0.5992  0.0558  0.1869 -0.4208
#>   0.5889 -0.0317  0.5741 -0.0554 -0.3576 -0.2083 -0.1908 -0.2552  0.0216
#>   0.4166  0.5877  0.2639 -0.0104 -0.1657  0.0408  0.3368  0.1481 -0.4174
#> 
#> Columns 10 to 18 -0.2244  0.2823 -0.6250  0.0687  0.1025 -0.0621  0.1072  0.5036  0.6797
#>   0.3023 -0.4427 -0.5568  0.0786 -0.6552  0.0820 -0.1434 -0.2814  0.5140
#>  -0.5998 -0.3700 -0.3938  0.3987 -0.7495 -0.4095  0.2248  0.1155  0.2680
#> 
#> Columns 19 to 20 -0.0596 -0.2973
#>  -0.4777  0.3691
#>  -0.2642  0.2653
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3321 -0.4241  0.0960  0.2421  0.4786 -0.1660  0.3607 -0.0814  0.0159
#>   0.3079 -0.5706  0.0987  0.0320 -0.3460 -0.0096 -0.3801 -0.5150  0.0832
#>   0.0519 -0.4476  0.3048 -0.3501 -0.2062  0.3452 -0.0830 -0.1818  0.3538
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0081  0.6220 -0.1637  0.2529  0.3615 -0.6857 -0.5326 -0.8165 -0.2657
#>  -0.4705 -0.3643 -0.0810  0.3840  0.3190 -0.2760  0.1447  0.0636  0.2184
#>   0.3998 -0.2537  0.2944  0.7772  0.5065  0.3447 -0.2731 -0.2003 -0.6247
#> 
#> Columns 10 to 18 -0.2773 -0.0145 -0.4527  0.5622  0.1360 -0.4833  0.2122 -0.2618  0.6869
#>  -0.3904 -0.4135 -0.2265  0.3006  0.3851 -0.4500 -0.3032  0.6804 -0.0140
#>  -0.2154 -0.4792 -0.3727  0.4320  0.3117  0.3942 -0.7883 -0.0253  0.3379
#> 
#> Columns 19 to 20  0.3976  0.4523
#>  -0.1912 -0.2262
#>  -0.2254 -0.8286
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3048 -0.5483  0.3059  0.2765 -0.2704  0.5296 -0.3930 -0.5296  0.3817
#>   0.2702 -0.6459 -0.1460  0.3629 -0.4018 -0.1516  0.2414 -0.4819  0.1875
#>   0.3453 -0.5309 -0.1411 -0.1985 -0.3338  0.2861  0.2988 -0.3766  0.1681
#> 
#> Columns 10 to 18  0.1580 -0.1161 -0.4867 -0.4488  0.0766  0.4727  0.1763 -0.4862  0.2130
#>   0.3672 -0.3924 -0.1186 -0.2266  0.4095  0.2409  0.4252 -0.6623  0.7879
#>   0.0043 -0.6269 -0.5837 -0.3541 -0.3635  0.3975 -0.1274 -0.7058  0.2098
#> 
#> Columns 19 to 20  0.2958  0.6307
#>  -0.3259  0.1548
#>  -0.1686  0.1917
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>