RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6755 -0.0243 -0.6819  0.8328  0.2549  0.3003 -0.1580 -0.0778  0.0202
#>  -0.3950  0.7486 -0.0212  0.3393  0.5310  0.2837 -0.0625 -0.2801  0.4187
#>  -0.4168 -0.0835 -0.1228  0.3043 -0.1076 -0.6991 -0.0715  0.2712 -0.1580
#> 
#> Columns 10 to 18  0.3625  0.3025 -0.2342  0.7743 -0.7343 -0.4726  0.5611 -0.3112  0.7945
#>   0.7291 -0.1783 -0.1416  0.5583 -0.6773 -0.8610 -0.5249 -0.6514  0.2466
#>  -0.1064 -0.5894  0.1283 -0.0220  0.0227 -0.5631 -0.0276 -0.3965 -0.6780
#> 
#> Columns 19 to 20 -0.5962  0.6485
#>  -0.4722 -0.2894
#>   0.4563 -0.3232
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.8286  0.1621 -0.5723  0.7362  0.2662 -0.3790  0.3024 -0.0556  0.0390
#>   0.5612  0.3273  0.1764  0.6614  0.0960  0.1900  0.1617 -0.0185  0.3103
#>  -0.1093  0.1712 -0.5616  0.0693 -0.1560 -0.5630  0.2708  0.0092 -0.1117
#> 
#> Columns 10 to 18  0.6483 -0.3479  0.1398  0.7262 -0.2911 -0.5509  0.0525  0.0555  0.2072
#>   0.5438 -0.1365  0.2735  0.6384 -0.3723 -0.6956 -0.2597 -0.1246  0.3558
#>   0.2865 -0.3199  0.2778  0.4491 -0.0925  0.0043  0.1488 -0.4729  0.4241
#> 
#> Columns 19 to 20  0.0176  0.2755
#>   0.0494 -0.0974
#>  -0.6485  0.1769
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.6173  0.2238 -0.3623 -0.0126  0.1716  0.2442 -0.2447  0.0022 -0.0419
#>   0.7758 -0.0047 -0.3044  0.8485  0.6531  0.2135  0.3573  0.0551 -0.0078
#>   0.6698  0.3519 -0.2764  0.1996 -0.1382 -0.2057 -0.0282 -0.0650  0.4259
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1922  0.2048 -0.6327  0.4514  0.3908  0.8188 -0.0884  0.1263  0.6877
#>   0.4268  0.2137  0.3151  0.4163 -0.6654 -0.0613 -0.3438  0.6942  0.6553
#>   0.3345 -0.1322 -0.3190 -0.0170  0.3288 -0.3636  0.4468  0.3483  0.1296
#> 
#> Columns 10 to 18  0.1912  0.0929 -0.6924 -0.4267  0.6382 -0.7565 -0.0696 -0.7230  0.3015
#>   0.6167 -0.7636 -0.0554  0.2480 -0.6230  0.6523 -0.3856 -0.3035  0.1402
#>  -0.5874  0.0616 -0.4097 -0.8245  0.1308 -0.0452 -0.1973  0.4483  0.1677
#> 
#> Columns 19 to 20  0.5151  0.1864
#>   0.2736  0.3353
#>   0.5004 -0.0588
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6650 -0.4321  0.2102  0.6376 -0.1048  0.4577  0.5228 -0.6598  0.0440
#>   0.5659 -0.3662  0.0722 -0.1202  0.1563  0.4312 -0.1840 -0.2048  0.3343
#>   0.7591  0.1728 -0.1444  0.2576  0.0639  0.1129  0.0378  0.2857 -0.0688
#> 
#> Columns 10 to 18 -0.6127 -0.3373  0.6653  0.5731 -0.1790  0.1881  0.3432 -0.4536 -0.0847
#>  -0.2115 -0.4088  0.2194  0.0992 -0.5171  0.0646  0.0215  0.1863  0.1001
#>  -0.2273 -0.5187  0.1633  0.4776 -0.3261 -0.2020  0.0572  0.1254 -0.3490
#> 
#> Columns 19 to 20  0.2186  0.6239
#>   0.3903  0.0565
#>  -0.1001 -0.0484
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>