RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8211  0.2414 -0.2881 -0.6408  0.4238 -0.1757  0.1385  0.0401 -0.4423
#>   0.7249  0.1904 -0.4617 -0.0694  0.3892 -0.9559  0.0686  0.1057  0.0291
#>   0.5062  0.0815  0.9037 -0.0264  0.5660  0.5089  0.6983 -0.7333 -0.5547
#> 
#> Columns 10 to 18  0.4854 -0.7086 -0.2461  0.7944 -0.2098  0.3661 -0.8907 -0.9220 -0.3205
#>   0.7588 -0.1247 -0.6977  0.1457  0.0592  0.6976  0.8694  0.6862 -0.2488
#>  -0.8485  0.0811  0.4940 -0.4814 -0.9258  0.0301  0.2266 -0.1842  0.7132
#> 
#> Columns 19 to 20  0.5664  0.4162
#>  -0.0856  0.8127
#>  -0.9800 -0.1670
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5213 -0.0236  0.6327 -0.1165 -0.6792  0.5468 -0.0758 -0.3318  0.3157
#>   0.0454  0.5226  0.4609 -0.2509  0.0838 -0.3635  0.3844 -0.2846 -0.4871
#>   0.4361 -0.1420  0.4323 -0.3181 -0.1505 -0.4595 -0.2887 -0.0003 -0.6695
#> 
#> Columns 10 to 18 -0.2157 -0.5504  0.5973  0.3285  0.4752 -0.5145  0.6794 -0.2632  0.0896
#>   0.2862 -0.6122  0.3216  0.3833 -0.2395  0.3962 -0.0915 -0.6240 -0.0211
#>  -0.2798 -0.0255  0.4228  0.0733 -0.5184 -0.0079  0.7490 -0.1489  0.5851
#> 
#> Columns 19 to 20  0.1161 -0.2820
#>   0.4613 -0.0011
#>  -0.7768 -0.4064
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1285 -0.0136  0.4441  0.0165  0.1862 -0.5954 -0.3366 -0.6034 -0.1043
#>  -0.3785 -0.5633  0.0866 -0.3909 -0.3297 -0.0416 -0.0308 -0.4052 -0.3496
#>  -0.1789  0.0316  0.0326 -0.3506  0.0252 -0.3687 -0.2313 -0.2290 -0.4631
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3030  0.4445  0.0461  0.3841  0.5295 -0.2875  0.1752 -0.0591 -0.2961
#>  -0.0611 -0.1591 -0.3449  0.2741  0.6622 -0.6057  0.2825 -0.0822 -0.0809
#>   0.6436  0.4594 -0.3725  0.7655 -0.0977  0.5290 -0.1161  0.8615 -0.2421
#> 
#> Columns 10 to 18  0.8211 -0.9335  0.9027 -0.8214 -0.6800 -0.4835  0.2024 -0.3933  0.0988
#>   0.7985 -0.1525  0.1467  0.1868  0.1360  0.2726  0.2120 -0.5277  0.1708
#>   0.2801 -0.3236 -0.8575  0.7054  0.0779 -0.4355  0.6122 -0.4037 -0.2720
#> 
#> Columns 19 to 20  0.6721 -0.1579
#>  -0.5763  0.2731
#>   0.0028  0.6740
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5509  0.4129  0.0843 -0.5840  0.5954  0.0718 -0.5124 -0.0754 -0.4080
#>   0.2865 -0.1189  0.4013 -0.5364  0.3391 -0.0973  0.0054 -0.1108 -0.5875
#>   0.5426  0.2944  0.5356 -0.3770 -0.3946  0.2650 -0.6453 -0.4339 -0.2752
#> 
#> Columns 10 to 18 -0.1275 -0.5453 -0.0057  0.1931 -0.4552  0.3086  0.6160 -0.3951  0.4433
#>  -0.0101 -0.4356  0.1925  0.1276 -0.5891  0.3317  0.5474 -0.4344  0.3365
#>  -0.1055 -0.1390  0.1393  0.2138  0.1732  0.3367  0.6429 -0.2744  0.6656
#> 
#> Columns 19 to 20 -0.3888 -0.1294
#>  -0.5974  0.2963
#>  -0.4768  0.2236
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>