RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6174 -0.6283  0.1804 -0.3659  0.1247  0.3812 -0.1204  0.3943  0.3656
#>   0.1474  0.5772 -0.4157  0.6902 -0.1069 -0.2698  0.5062  0.0318 -0.6403
#>  -0.0822 -0.1599  0.1981  0.3724  0.4558 -0.2436 -0.1946 -0.4146 -0.2041
#> 
#> Columns 10 to 18  0.0928  0.7597 -0.6311 -0.7723 -0.3043  0.7867 -0.8325  0.0964 -0.3309
#>  -0.3683  0.0790 -0.1348  0.4349 -0.0262  0.6240 -0.7006  0.5693  0.4698
#>   0.2877 -0.0843  0.4903 -0.0985 -0.0511  0.7348 -0.1077  0.6234  0.1397
#> 
#> Columns 19 to 20 -0.0227  0.8895
#>  -0.2008  0.7770
#>  -0.8478 -0.0400
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1022 -0.1448 -0.0525  0.4538 -0.0955 -0.1786 -0.0998  0.0153 -0.0424
#>  -0.5145  0.0379 -0.1035  0.6122 -0.2604  0.2237  0.2682 -0.2199 -0.2865
#>  -0.2462 -0.0115 -0.3237  0.4001 -0.6295  0.1778  0.2656 -0.3103 -0.0857
#> 
#> Columns 10 to 18 -0.0799 -0.2831  0.2107  0.3228 -0.1536  0.3104  0.1248  0.1951 -0.1624
#>   0.3734  0.4805  0.3095 -0.2760 -0.3919  0.2608 -0.4244  0.3968  0.5153
#>  -0.0322  0.2065 -0.1394 -0.3813 -0.4906  0.0398 -0.4106 -0.1227  0.4411
#> 
#> Columns 19 to 20  0.0347 -0.1256
#>  -0.2827  0.0213
#>  -0.1216  0.6511
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1586 -0.4119 -0.1531  0.5591 -0.1124 -0.0263  0.0067 -0.1524  0.0624
#>  -0.3569 -0.2601  0.0590 -0.0586  0.2094  0.2269 -0.0442  0.0964 -0.1925
#>  -0.5310 -0.1948  0.2357  0.3831  0.3559 -0.3031 -0.2543  0.2595 -0.1768
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3874 -0.6027  0.2305  0.6754  0.7179  0.2323 -0.4139  0.7262  0.5022
#>   0.4407 -0.3565 -0.1145 -0.3063 -0.4486  0.3873  0.1592 -0.1757  0.4274
#>  -0.4963 -0.1802  0.5900  0.7666  0.0740  0.0533 -0.0083  0.1752  0.2239
#> 
#> Columns 10 to 18  0.5382  0.5488  0.6869 -0.4856  0.2824 -0.7486 -0.2664  0.2116  0.4591
#>   0.3469 -0.3763 -0.3778  0.3088 -0.5194 -0.0182 -0.8118  0.0715 -0.0497
#>   0.2398  0.1536  0.5945  0.4042  0.2533  0.3438  0.0222  0.3622  0.7477
#> 
#> Columns 19 to 20  0.2822 -0.0585
#>  -0.2337  0.1842
#>   0.5817  0.0807
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3077 -0.3816 -0.7003  0.3431  0.1155 -0.4605  0.2093 -0.2509  0.0477
#>  -0.2538 -0.1660  0.0333  0.3489 -0.0250 -0.1381  0.1900 -0.0990  0.3344
#>  -0.3645 -0.4338 -0.0352  0.3937  0.3571 -0.1446 -0.0686  0.0927 -0.1681
#> 
#> Columns 10 to 18 -0.2776 -0.2667  0.6359 -0.2188 -0.2015 -0.0001  0.4910  0.0632  0.2829
#>   0.2329  0.3817 -0.0734 -0.3577 -0.2968  0.1627 -0.3571  0.1134  0.4060
#>  -0.1087  0.3604  0.4111 -0.0784 -0.0760  0.3613 -0.3854  0.4061  0.3499
#> 
#> Columns 19 to 20 -0.2298  0.1199
#>  -0.4656  0.6869
#>  -0.4504  0.5944
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>