RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4844 -0.3965  0.0796 -0.7495  0.0825 -0.2792 -0.5622  0.3938 -0.5562
#>   0.8834  0.3682 -0.6811  0.5006 -0.3613 -0.8836 -0.1058  0.0790  0.5039
#>  -0.0283  0.6321  0.8547 -0.5688 -0.5271  0.2915  0.0577 -0.9416 -0.8653
#> 
#> Columns 10 to 18 -0.7545  0.0250 -0.0147 -0.4759  0.3892  0.7977 -0.0318 -0.1624  0.8030
#>   0.9513  0.3652  0.0756  0.2955 -0.6889 -0.2031 -0.4592  0.6075  0.5922
#>   0.1504 -0.3706 -0.7479  0.7255  0.3798  0.2860 -0.4298 -0.8240 -0.2710
#> 
#> Columns 19 to 20 -0.0011 -0.5219
#>  -0.2128  0.1017
#>  -0.3243  0.3936
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2880 -0.0610  0.0313  0.0265 -0.1109 -0.1891 -0.1699 -0.5099  0.2234
#>   0.1110  0.5227  0.4584 -0.6365  0.2308  0.1754 -0.2521 -0.2982 -0.4544
#>  -0.0832  0.3111  0.0766  0.6464  0.0523 -0.5522 -0.3051 -0.4360 -0.1728
#> 
#> Columns 10 to 18 -0.6322  0.1817  0.2630  0.5345  0.2255 -0.4719  0.2384  0.1573  0.1531
#>  -0.2605  0.0146  0.4078 -0.1833  0.1589  0.4391 -0.2925  0.0109  0.1491
#>   0.1989 -0.2557 -0.2216 -0.1384 -0.1135 -0.5185  0.0321 -0.0115  0.4843
#> 
#> Columns 19 to 20 -0.7144  0.5273
#>  -0.1721 -0.1079
#>   0.3416  0.2251
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2420  0.3100  0.4542  0.1094  0.4192 -0.0079 -0.4134 -0.2275 -0.0426
#>   0.3840 -0.1671  0.0378  0.0606  0.1491 -0.1201 -0.4029 -0.4633  0.1151
#>  -0.4248  0.2080  0.3962 -0.1806  0.3789 -0.5571 -0.4471 -0.3555 -0.2936
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4150  0.2371  0.6392  0.5974  0.4684 -0.1704 -0.4064 -0.1411 -0.7460
#>   0.1978  0.2177  0.2864 -0.4140 -0.1480 -0.5136 -0.0859 -0.6165 -0.4357
#>   0.3469  0.5199 -0.6028  0.2730  0.3584  0.1757 -0.1551 -0.0981  0.1592
#> 
#> Columns 10 to 18 -0.3877  0.3137  0.5246 -0.1277  0.2250  0.2863  0.2350  0.2874 -0.5932
#>   0.1670  0.5323  0.6033 -0.8668 -0.6610 -0.2939 -0.0344  0.1554 -0.6278
#>   0.0137  0.2176 -0.2281 -0.2808 -0.4176  0.4610  0.2519  0.0560  0.5113
#> 
#> Columns 19 to 20 -0.3857 -0.2282
#>  -0.6420  0.5092
#>  -0.0374  0.4974
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0470  0.3068  0.0605 -0.0163  0.2077  0.2523 -0.2664 -0.3914 -0.0825
#>   0.0718 -0.1380  0.1969 -0.3605  0.0496 -0.1337 -0.2371 -0.4238  0.0454
#>  -0.4585 -0.2980  0.0897  0.2721  0.4701 -0.4721 -0.4881 -0.3268  0.1986
#> 
#> Columns 10 to 18  0.0784  0.0994  0.0374  0.5640  0.3371 -0.0468 -0.4608 -0.0814 -0.2620
#>  -0.7301  0.6596  0.2814  0.2393  0.2379  0.1803 -0.1417 -0.0372  0.1851
#>  -0.0761  0.0676  0.2942  0.4283  0.0607  0.2280 -0.0881 -0.3423  0.1842
#> 
#> Columns 19 to 20 -0.3880  0.5484
#>   0.0512  0.6889
#>  -0.3574  0.0548
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>