RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7283  0.9311 -0.7173 -0.8273 -0.3193 -0.1927 -0.0097  0.1461  0.2102
#>   0.1805 -0.2610 -0.2199  0.3724 -0.7157  0.2894  0.3480 -0.8186  0.2305
#>  -0.2473  0.1269 -0.0218  0.2696  0.1616  0.6343 -0.5507  0.0086  0.5819
#> 
#> Columns 10 to 18 -0.3407  0.2165 -0.9246 -0.2189  0.2861  0.6914 -0.7376  0.1887  0.7874
#>   0.7243 -0.2787 -0.4628 -0.4759  0.1985 -0.0053 -0.0804 -0.3312  0.3895
#>  -0.4766 -0.2879  0.1857  0.5584  0.0160  0.4110 -0.0786 -0.1701 -0.0737
#> 
#> Columns 19 to 20  0.3121 -0.5030
#>   0.8205 -0.3202
#>   0.0106  0.1669
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0450  0.0505  0.0665  0.4282 -0.2347  0.9109  0.3190 -0.5732  0.0840
#>  -0.4147  0.2958  0.0464  0.1644 -0.3058  0.2424 -0.1411 -0.0084  0.0236
#>  -0.6894 -0.3669  0.4522  0.4840 -0.4760  0.5284 -0.5730 -0.4949 -0.1552
#> 
#> Columns 10 to 18  0.3725 -0.6839 -0.1619 -0.1249 -0.3990 -0.0677  0.4156 -0.6017  0.2033
#>  -0.0241 -0.0665 -0.3131  0.3427  0.3082 -0.6809  0.3840  0.1367  0.6466
#>   0.4108  0.0641  0.2187  0.4255  0.0072 -0.1447  0.2385  0.4449  0.7304
#> 
#> Columns 19 to 20 -0.4893 -0.1031
#>  -0.5184  0.0427
#>  -0.4586  0.3754
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0789  0.4426  0.2704 -0.0132  0.4040  0.1517 -0.3569 -0.4914  0.4572
#>  -0.2073  0.3006  0.1473  0.6514  0.1567  0.4716  0.1498 -0.4798  0.4469
#>   0.3255  0.3561  0.4360  0.2474  0.1913  0.3695 -0.1105  0.2904  0.4149
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1640 -0.6212 -0.7513 -0.7217 -0.1991 -0.5169  0.7145 -0.7398 -0.6005
#>  -0.4054 -0.5711  0.1006 -0.7924 -0.5120 -0.1529  0.0918 -0.0544  0.3028
#>   0.0347 -0.5652 -0.3099 -0.5379  0.4061 -0.2501  0.6380 -0.3115 -0.4665
#> 
#> Columns 10 to 18 -0.4460 -0.0359  0.4328  0.6916 -0.5671  0.0030 -0.3571 -0.5911 -0.2670
#>   0.0053  0.1187 -0.6139  0.3233  0.5215  0.2377  0.3032  0.1015  0.0330
#>  -0.4629  0.3371  0.3358  0.0126 -0.4757 -0.5079 -0.3756 -0.6226 -0.1002
#> 
#> Columns 19 to 20  0.2451  0.1297
#>   0.1785  0.0513
#>  -0.4124  0.0442
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1430  0.3914  0.1583  0.6046 -0.3378  0.6867  0.0377 -0.0851  0.0967
#>  -0.0572  0.3396  0.4010  0.0408  0.0295  0.3742 -0.2412 -0.0871  0.3009
#>  -0.1272  0.6642 -0.1417  0.4638  0.0911  0.2502 -0.1165 -0.0367 -0.0060
#> 
#> Columns 10 to 18  0.0734 -0.2734  0.0091  0.6932 -0.0661 -0.6793  0.5257  0.4248  0.6852
#>   0.2853 -0.2646  0.0592  0.4051 -0.2128 -0.4927  0.1869  0.1158  0.3334
#>   0.1006 -0.0330  0.2396  0.3898  0.0250 -0.4182  0.2196  0.5559  0.4849
#> 
#> Columns 19 to 20 -0.5727 -0.3398
#>  -0.3621 -0.2469
#>  -0.6202  0.2014
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>