RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6143 -0.5959 -0.0268  0.3317 -0.1526  0.1492 -0.2749  0.7742  0.2507
#>  -0.0998 -0.9404  0.4495  0.5202 -0.4414  0.2222 -0.4008  0.4032  0.2007
#>  -0.4702  0.5833  0.1502  0.8538 -0.5935 -0.2829 -0.5471 -0.5777  0.6538
#> 
#> Columns 10 to 18  0.2332  0.2138  0.3025 -0.3067 -0.4798  0.0071 -0.8103 -0.8250  0.8483
#>   0.2707  0.5048 -0.0413 -0.2443 -0.7352  0.2336 -0.9254 -0.1866  0.7731
#>  -0.5043  0.6857  0.5629  0.7045  0.9535  0.8047 -0.7015 -0.7243  0.4240
#> 
#> Columns 19 to 20  0.1414  0.6028
#>  -0.5017 -0.3660
#>  -0.4132  0.9414
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5856 -0.7293  0.0766  0.1741 -0.2862  0.1921 -0.0384  0.3301  0.4534
#>  -0.5865 -0.4197 -0.6329 -0.2580 -0.2290 -0.0770  0.1149  0.1168  0.2690
#>  -0.1073 -0.4384 -0.2174  0.6364 -0.8235  0.1945  0.8170 -0.2610  0.5736
#> 
#> Columns 10 to 18  0.0118 -0.0426 -0.0640  0.3302 -0.1771 -0.3879 -0.6095  0.2775  0.3505
#>  -0.2816  0.0839 -0.1314  0.0677 -0.1290  0.7660  0.1366  0.4999 -0.0814
#>  -0.6453 -0.0775 -0.4766  0.6016 -0.0703  0.5037 -0.1112  0.0884 -0.6805
#> 
#> Columns 19 to 20  0.2942  0.5400
#>   0.4300  0.5423
#>   0.6239 -0.1587
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2485 -0.1516 -0.1471  0.4884  0.2980  0.0359 -0.0633 -0.4276  0.7460
#>  -0.1063  0.1976  0.0751  0.4332 -0.0127 -0.1588  0.1995  0.5257  0.4447
#>  -0.0048  0.5163 -0.4907  0.7931 -0.3960  0.1857  0.3275  0.1929  0.2667
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2118  0.6427  0.2262  0.3529 -0.4667  0.4819 -0.6513 -0.5601 -0.6577
#>  -0.6995 -0.7489 -0.2587  0.0340  0.7919  0.7495 -0.2240  0.1809  0.3590
#>  -0.1735  0.2341 -0.1526  0.4759  0.4839  0.6168 -0.3604 -0.0965 -0.2589
#> 
#> Columns 10 to 18  0.6476  0.1076  0.1422 -0.2537 -0.2643 -0.1459  0.0575 -0.3064  0.7373
#>   0.3937 -0.5726  0.1513  0.6580  0.6831 -0.7449  0.1475  0.5937  0.1383
#>   0.4925 -0.0605  0.4249 -0.0474  0.3736 -0.1284 -0.1974  0.3315 -0.0406
#> 
#> Columns 19 to 20 -0.5487  0.6137
#>   0.6184 -0.2205
#>   0.5596  0.6528
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2714 -0.3056 -0.0753  0.1805 -0.2525  0.1314  0.6137 -0.0221  0.1702
#>   0.0284 -0.6269  0.2058  0.8155  0.3736  0.0447  0.1048 -0.5762  0.5421
#>  -0.1305 -0.2300  0.0976  0.7124 -0.0161  0.0368  0.2391 -0.0302  0.5409
#> 
#> Columns 10 to 18 -0.1659  0.0050  0.1296  0.2465  0.1315  0.2136 -0.0381  0.0943 -0.3887
#>   0.2594  0.7320  0.0199 -0.5361 -0.5213  0.0462 -0.4652 -0.4813  0.7172
#>   0.0316  0.4398 -0.3637  0.0334 -0.3789  0.0359 -0.4234  0.0250  0.3766
#> 
#> Columns 19 to 20  0.6509  0.6344
#>  -0.4092  0.6380
#>  -0.0618  0.5406
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>