RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5431 -0.1337  0.0442  0.0526 -0.4647  0.1466 -0.0918 -0.2144  0.4667
#>  -0.1414 -0.1861  0.4619  0.7146 -0.0006 -0.2816 -0.4675  0.2505  0.3869
#>   0.7353 -0.1449  0.2415  0.9496  0.1242  0.1346  0.1947  0.3440  0.0831
#> 
#> Columns 10 to 18 -0.6299  0.4171  0.2894 -0.4664  0.3104 -0.1046  0.6337 -0.4567 -0.4380
#>  -0.5749  0.3910 -0.2448  0.7520  0.5133  0.2675 -0.1357  0.4437  0.7370
#>   0.3814  0.0350  0.1502  0.0090  0.1284 -0.1562  0.5726  0.7460  0.7444
#> 
#> Columns 19 to 20  0.2727 -0.3798
#>   0.2733 -0.1709
#>  -0.5867  0.7306
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1251 -0.6201 -0.0753 -0.1582 -0.1037 -0.2234 -0.3000  0.2132  0.2199
#>   0.2391 -0.7290 -0.0825  0.6472 -0.3667  0.1396 -0.3222  0.0774  0.5197
#>   0.1106 -0.6755 -0.1222  0.5187 -0.3511  0.4279  0.1279  0.5109  0.2198
#> 
#> Columns 10 to 18 -0.4815  0.5262 -0.1961 -0.6306  0.2105  0.0270  0.2838  0.1905 -0.4133
#>  -0.3897  0.4051 -0.0366 -0.0761  0.4148 -0.3584 -0.1463 -0.0768  0.2097
#>  -0.0434  0.5173  0.2562 -0.3335  0.0022  0.0264  0.2141  0.6277  0.2508
#> 
#> Columns 19 to 20 -0.2013 -0.1467
#>  -0.5935 -0.1409
#>  -0.4807  0.2749
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2212 -0.8668  0.0857  0.3853 -0.0708 -0.0965  0.1664  0.0046  0.5617
#>   0.1641 -0.6954 -0.1137  0.6338 -0.2272  0.1963 -0.3316  0.0414  0.5704
#>  -0.4462 -0.6028 -0.5800  0.3513  0.0579  0.0687  0.2786 -0.2704 -0.1411
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1149  0.5081 -0.6912 -0.3013 -0.2427 -0.4951  0.5076  0.1186  0.4717
#>   0.6965 -0.4776 -0.1954 -0.2613  0.2883  0.3866 -0.6516  0.2144  0.4470
#>  -0.6129  0.4595 -0.7694 -0.3018  0.2566 -0.5779 -0.5476 -0.1203 -0.2045
#> 
#> Columns 10 to 18 -0.1834  0.5647  0.4481 -0.3068  0.3356  0.3464  0.4730 -0.5025 -0.1977
#>  -0.1141  0.4382  0.5462  0.2445 -0.8533  0.8076  0.4042 -0.0830  0.3522
#>   0.0785  0.3792  0.2902  0.6177  0.1975  0.2652  0.0048 -0.6565 -0.3430
#> 
#> Columns 19 to 20 -0.3908  0.6064
#>  -0.0768 -0.2026
#>  -0.7324  0.1776
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2343 -0.6447  0.4850  0.6104 -0.5000  0.0221  0.2806  0.4991  0.5019
#>   0.5355 -0.4334  0.2466  0.4397  0.0947 -0.1796 -0.2102 -0.0999  0.4611
#>   0.0694 -0.7276  0.0522  0.5001 -0.2998  0.1158  0.2594  0.2544  0.4958
#> 
#> Columns 10 to 18 -0.4551  0.4937  0.0807 -0.2335  0.3000 -0.3710  0.0906  0.4008  0.1247
#>  -0.1113  0.5729  0.1341  0.0713 -0.1273 -0.3236  0.0675 -0.1454  0.0066
#>   0.1188  0.2672  0.3570 -0.5274 -0.1104 -0.0472  0.0450  0.5302  0.0901
#> 
#> Columns 19 to 20 -0.5716 -0.0491
#>  -0.4295  0.1251
#>  -0.5213  0.0241
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>