RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0555  0.8708 -0.4326 -0.5129  0.3506  0.4099  0.4612 -0.7562 -0.2083
#>  -0.7285 -0.0010 -0.2743 -0.4015 -0.6251  0.3409 -0.0012 -0.0393 -0.6537
#>  -0.1563 -0.4061  0.1994  0.8062 -0.7327  0.4287 -0.6747  0.0965  0.5649
#> 
#> Columns 10 to 18 -0.2898 -0.4364 -0.1834  0.2629 -0.8755 -0.6798 -0.5405  0.6457 -0.5048
#>   0.7285  0.9668  0.8690  0.5844 -0.1338  0.3138  0.3272  0.0510 -0.4901
#>   0.5900  0.7669  0.2147 -0.5490 -0.1991 -0.5884  0.3379 -0.7765 -0.5197
#> 
#> Columns 19 to 20 -0.0768  0.5256
#>   0.3725 -0.8987
#>   0.3904  0.6203
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1799  0.0793 -0.0187  0.5450  0.3441 -0.4671 -0.0246  0.0595  0.5203
#>  -0.6476  0.6303  0.0386 -0.3011  0.2483 -0.4772  0.1614 -0.5896  0.5651
#>  -0.4491  0.5067 -0.2236 -0.3771 -0.0972 -0.4551  0.0013  0.1619  0.1746
#> 
#> Columns 10 to 18  0.3199 -0.3610  0.2830 -0.2426  0.1800 -0.6728  0.5401 -0.4491  0.4595
#>  -0.5603  0.1483  0.4942  0.1690  0.1091  0.0819  0.5270 -0.4910 -0.0752
#>  -0.5010  0.0694  0.2290  0.0520 -0.6692 -0.4311  0.7262  0.3432 -0.5220
#> 
#> Columns 19 to 20  0.1693 -0.6301
#>  -0.0118 -0.3810
#>   0.2377  0.0887
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4846 -0.0447 -0.0019 -0.2636  0.0061  0.3571  0.0139 -0.2980 -0.2732
#>  -0.4767 -0.3858 -0.0300  0.6247 -0.1609 -0.2805 -0.2095  0.0733 -0.0914
#>  -0.3955 -0.1759 -0.2791  0.4859 -0.4792 -0.6347 -0.1523  0.3764  0.5080
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2489  0.2567  0.0457 -0.8820 -0.4092 -0.5167 -0.1655  0.1903 -0.2687
#>  -0.2735 -0.1633  0.3782 -0.4294 -0.1705 -0.4009  0.4670 -0.2505  0.8230
#>   0.6108 -0.3632  0.6463 -0.7851 -0.6273 -0.4405  0.2607  0.2902  0.0098
#> 
#> Columns 10 to 18  0.6096  0.3611 -0.3484 -0.7691 -0.4811 -0.5185  0.2450  0.4369 -0.3407
#>   0.0465 -0.4735 -0.0246 -0.5319 -0.3800 -0.4084  0.3467  0.4341 -0.5138
#>  -0.3351 -0.2275  0.2891 -0.4564 -0.3479 -0.5362  0.4310  0.6729 -0.5780
#> 
#> Columns 19 to 20 -0.5251 -0.6130
#>  -0.0566  0.2066
#>  -0.6082 -0.5804
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4940  0.3147  0.1331 -0.2013  0.0101 -0.1119  0.2294 -0.3075  0.2432
#>  -0.3203 -0.0164 -0.0294  0.0886  0.0415 -0.4520 -0.5812  0.0072  0.4961
#>  -0.0013 -0.0156  0.1490 -0.1482  0.1228 -0.1504 -0.1156 -0.1874  0.1937
#> 
#> Columns 10 to 18 -0.2075 -0.0705  0.2819  0.2052  0.1297 -0.2391  0.5880 -0.3276  0.1540
#>   0.1059  0.1265  0.5466  0.0969 -0.1701 -0.1750  0.5084 -0.4272 -0.0404
#>   0.3684  0.2677  0.5759 -0.0293 -0.0318 -0.3582  0.2133 -0.2474  0.2122
#> 
#> Columns 19 to 20 -0.3682 -0.2078
#>   0.1857 -0.6195
#>  -0.3482 -0.3427
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>