RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7825  0.2045 -0.2926 -0.4013 -0.4648  0.5122  0.8864 -0.6923  0.0632
#>  -0.4971 -0.3565  0.4452 -0.3993 -0.3390 -0.6344  0.6416  0.3174  0.1549
#>  -0.3637 -0.7461  0.4377 -0.3141  0.8379 -0.9208 -0.1591  0.6940  0.3075
#> 
#> Columns 10 to 18  0.0041 -0.6886  0.5679  0.6477  0.6249  0.3401 -0.8697 -0.9572  0.6290
#>   0.4520 -0.5774  0.6888  0.7950  0.1424  0.3845  0.4259  0.4616  0.6575
#>  -0.0773  0.1111  0.8758 -0.0279  0.5169  0.1082 -0.7736  0.7513 -0.6428
#> 
#> Columns 19 to 20 -0.4788 -0.5430
#>   0.5122 -0.5102
#>   0.0248 -0.3084
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3417  0.3344  0.7331 -0.1412 -0.0782 -0.0267  0.0054 -0.3292  0.0963
#>   0.4569 -0.3391  0.0594 -0.0904 -0.1798 -0.3576 -0.1796  0.2115 -0.5400
#>   0.1508 -0.3678 -0.2971  0.3624 -0.2861 -0.7593  0.2243  0.1882  0.3711
#> 
#> Columns 10 to 18  0.2036 -0.3086 -0.2254  0.3845  0.0313  0.1949 -0.6621  0.3359 -0.5095
#>   0.4795 -0.1699  0.4599 -0.2752  0.2848 -0.4678 -0.2363 -0.4639 -0.6818
#>   0.2372 -0.0931  0.6454  0.1819  0.5053 -0.1714  0.3948 -0.4722  0.0590
#> 
#> Columns 19 to 20 -0.5653 -0.6353
#>   0.0714 -0.2319
#>   0.2114  0.3467
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0210 -0.0883  0.0185  0.4344  0.4352 -0.3499  0.3712 -0.1041 -0.0294
#>   0.2552  0.1485 -0.2369  0.5358  0.0087 -0.2183  0.6140 -0.6459 -0.3515
#>  -0.0379  0.1840 -0.0216  0.2923 -0.3884 -0.3803  0.1978 -0.5351  0.0432
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6751 -0.8361  0.5553  0.3715 -0.7771 -0.4527  0.2925 -0.2307  0.5655
#>   0.0608 -0.1318 -0.2810  0.1265 -0.7518 -0.1471  0.6337 -0.6258 -0.3168
#>   0.5960  0.8160 -0.2774 -0.0058  0.3303  0.3405 -0.5550 -0.4629 -0.6662
#> 
#> Columns 10 to 18 -0.3592 -0.2073  0.5519  0.7315 -0.5524  0.0007 -0.4035 -0.5054  0.1704
#>  -0.8269  0.0208  0.1445  0.9025 -0.8348 -0.4831 -0.1766  0.3450 -0.2346
#>   0.8753  0.6845 -0.2428  0.5462 -0.5026  0.1452  0.4401  0.5193  0.5706
#> 
#> Columns 19 to 20  0.2301  0.1614
#>  -0.0539  0.4278
#>  -0.8001  0.7531
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1858 -0.0292 -0.5801 -0.0161 -0.3239 -0.2975  0.1281  0.1783 -0.7207
#>  -0.2451  0.3920 -0.5444 -0.0682 -0.4320 -0.4285  0.1047 -0.1649 -0.5535
#>  -0.5139 -0.2241  0.2511  0.3534 -0.6100 -0.1395 -0.5413  0.1590  0.4946
#> 
#> Columns 10 to 18  0.5319 -0.3966  0.1027  0.1913  0.3097 -0.3565 -0.4681 -0.3889  0.1707
#>   0.5679 -0.0214 -0.1920  0.4166 -0.2420 -0.5959  0.0408  0.3297 -0.5395
#>   0.6503  0.0452  0.0278  0.2094 -0.0731 -0.4353  0.1420  0.4008 -0.4732
#> 
#> Columns 19 to 20  0.6954 -0.3957
#>   0.4974 -0.2533
#>  -0.0872 -0.3676
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>