RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0899  0.8237 -0.5829 -0.5276 -0.3971  0.0356  0.6141  0.3537 -0.4992
#>   0.5494  0.2729 -0.5455 -0.4998 -0.5567  0.3240 -0.1017 -0.0705  0.0024
#>   0.6326  0.4088 -0.0428 -0.1734  0.3813 -0.1291 -0.5551  0.6194  0.2561
#> 
#> Columns 10 to 18  0.2277  0.2041 -0.6206 -0.1013 -0.2132  0.6093  0.2600 -0.8588 -0.9054
#>   0.0857  0.2128 -0.7542  0.6987 -0.5446  0.3889 -0.5264 -0.5705  0.6400
#>  -0.2127  0.7690 -0.2239  0.3827 -0.1835  0.8069  0.0004 -0.6186 -0.6185
#> 
#> Columns 19 to 20 -0.0528  0.3702
#>   0.5873  0.2026
#>   0.8262  0.6748
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1554 -0.0725  0.4271 -0.2584  0.1170 -0.3534  0.1340  0.2619 -0.1529
#>   0.0871  0.3154  0.0138 -0.4170  0.1806 -0.3398 -0.1533  0.0723 -0.1709
#>  -0.0155 -0.1705 -0.3555  0.2295  0.4048 -0.6029  0.0113  0.3022 -0.0763
#> 
#> Columns 10 to 18  0.0942  0.6150  0.1609 -0.6030 -0.3743  0.5365  0.0807 -0.2830 -0.1043
#>   0.4023 -0.0124 -0.1549  0.2750 -0.4608  0.0755 -0.2663  0.2878  0.3742
#>   0.5143  0.2438 -0.3881 -0.0224 -0.1902  0.0801 -0.4686  0.4142  0.1108
#> 
#> Columns 19 to 20  0.3041  0.0375
#>   0.5533  0.3162
#>   0.6164  0.0035
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2228 -0.0383  0.2209 -0.5495  0.3959 -0.4584 -0.3863  0.4867  0.1131
#>   0.3526  0.1499  0.2139 -0.2465 -0.0076 -0.1944 -0.1160  0.2657  0.2184
#>   0.2711  0.5986 -0.2676 -0.3094 -0.3079 -0.1201 -0.2512  0.3767  0.3082
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8560  0.4694  0.6206  0.3431  0.5469 -0.6956  0.1496  0.6809  0.9482
#>   0.1634  0.5996  0.1746 -0.3983  0.6343  0.0175  0.6047  0.3270  0.6999
#>  -0.1534  0.6012  0.2215  0.1831 -0.0873 -0.2766 -0.0181  0.5442  0.7670
#> 
#> Columns 10 to 18 -0.4073 -0.2171  0.3698  0.8094 -0.5680  0.1073  0.3947  0.4547 -0.6201
#>   0.0356 -0.4138 -0.4190 -0.1204 -0.5855 -0.0455  0.3346  0.3056 -0.5822
#>  -0.3222  0.0845 -0.1133  0.1346  0.2405 -0.0939  0.3126 -0.0265 -0.2151
#> 
#> Columns 19 to 20 -0.2582 -0.5804
#>  -0.7930 -0.0335
#>   0.2571  0.3447
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.7211 -0.6760  0.1400 -0.1088  0.1250 -0.0605 -0.3985  0.3647 -0.0291
#>   0.5934 -0.4111  0.3598 -0.4814  0.3908 -0.4948 -0.1072  0.3759  0.3938
#>   0.4499 -0.2727  0.2185 -0.5264  0.3603 -0.1793 -0.4422  0.1696  0.2814
#> 
#> Columns 10 to 18 -0.2260  0.3929 -0.2249  0.1811  0.1446 -0.1301  0.1743  0.2294  0.4014
#>   0.5871  0.3475  0.1132 -0.0524  0.1380  0.2700  0.7828 -0.0904  0.6017
#>   0.3487  0.3098 -0.2558  0.1678 -0.0366 -0.2457  0.4010 -0.1619  0.3191
#> 
#> Columns 19 to 20  0.5955 -0.5059
#>   0.6489 -0.1295
#>   0.5748 -0.1651
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>