RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4694 -0.8572 -0.0558  0.0687 -0.0898 -0.0848  0.8578 -0.0636 -0.8378
#>  -0.0845  0.8629 -0.4343  0.5670 -0.4483 -0.8910 -0.3225  0.7137 -0.2076
#>   0.1683  0.2857 -0.4295  0.3470  0.7098  0.7388 -0.2993  0.3974 -0.2303
#> 
#> Columns 10 to 18  0.3486 -0.2664  0.1451  0.6499  0.5696  0.1984  0.0705 -0.8134 -0.1996
#>   0.2572 -0.2791 -0.7749 -0.1410  0.1511 -0.5870 -0.1082  0.4510  0.4853
#>   0.1467 -0.7328 -0.5275  0.4664 -0.7752 -0.3361  0.7962 -0.9331 -0.6758
#> 
#> Columns 19 to 20  0.3363  0.4182
#>   0.9350 -0.1188
#>  -0.3297 -0.1765
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1527 -0.1781  0.1625  0.4731  0.3217  0.5637 -0.1019  0.5201  0.1286
#>   0.0355  0.5614  0.5038  0.1933  0.3631  0.1187  0.7524  0.0247 -0.3892
#>   0.3876  0.3140 -0.0833  0.0005  0.1245 -0.5137  0.5538 -0.4504 -0.4223
#> 
#> Columns 10 to 18 -0.7871 -0.6979 -0.1917  0.2057  0.1341 -0.3157  0.2184 -0.0078 -0.1829
#>  -0.3189 -0.1888 -0.2405  0.1874  0.0934  0.2216  0.4906 -0.2654 -0.0871
#>  -0.3241 -0.3332 -0.3038  0.5403 -0.2897 -0.2619  0.0751 -0.3852 -0.1600
#> 
#> Columns 19 to 20  0.3171  0.4255
#>   0.4938 -0.5395
#>   0.1598 -0.1585
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4331 -0.0243  0.2152  0.4639  0.3040  0.2542  0.8272 -0.4108 -0.5334
#>  -0.1879  0.1046  0.4828  0.4755 -0.1127  0.4603 -0.1860  0.3889  0.0664
#>   0.1509  0.2392  0.4931  0.1689 -0.1281  0.0701  0.2982  0.3360 -0.0678
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1438 -0.2607  0.7618 -0.1385  0.1930 -0.6344  0.3637  0.1426 -0.1733
#>  -0.2669 -0.2298 -0.0310  0.6441  0.1122  0.0571  0.0365 -0.2033 -0.1564
#>   0.2382 -0.0052 -0.1923  0.0967  0.5397 -0.1045  0.6079 -0.5682 -0.3058
#> 
#> Columns 10 to 18  0.1051 -0.3860  0.1449  0.5394  0.0198 -0.1382  0.1613  0.2858 -0.5918
#>   0.2200 -0.3683 -0.3152  0.4424 -0.6074 -0.7598 -0.3141 -0.0581 -0.4951
#>  -0.6209 -0.3809 -0.5257  0.2357 -0.1816 -0.7471 -0.5268 -0.2496  0.0035
#> 
#> Columns 19 to 20  0.5560 -0.4462
#>  -0.1895 -0.1224
#>  -0.1897 -0.4578
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4360  0.2430  0.4184  0.1894  0.1552  0.2580  0.3648  0.0977 -0.3307
#>   0.5724  0.2276  0.2923  0.1075 -0.2290  0.5137  0.5345  0.4702 -0.3372
#>   0.3141  0.1204  0.2308  0.2632  0.0859  0.6580  0.2773  0.4221  0.1804
#> 
#> Columns 10 to 18 -0.0150  0.0860 -0.0616 -0.2369 -0.3643  0.1263 -0.3056 -0.4474  0.0699
#>  -0.1267 -0.2644 -0.4215 -0.3313  0.1844 -0.1261 -0.0501 -0.4252 -0.3007
#>  -0.2117 -0.1376 -0.2411 -0.4359  0.1281  0.0271  0.1652 -0.3339 -0.1849
#> 
#> Columns 19 to 20  0.7173 -0.2183
#>   0.4389  0.1617
#>   0.6136  0.4004
#> [ CPUFloatType{2,3,20} ]
#>