RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8220  0.4772 -0.5583  0.3481 -0.7328 -0.0959 -0.0776 -0.2644  0.6639
#>   0.4043 -0.6729  0.2919  0.5031 -0.2950 -0.3355 -0.2693 -0.4637  0.2950
#>  -0.2092 -0.2928 -0.1287 -0.7836  0.0989 -0.7921  0.3763  0.2321 -0.1178
#> 
#> Columns 10 to 18 -0.1879 -0.0732 -0.4161  0.5158  0.8872 -0.0880 -0.4133 -0.3868 -0.4263
#>  -0.8990 -0.8107 -0.2743  0.6425 -0.5307 -0.2887  0.6613 -0.7098 -0.0108
#>   0.3368  0.7416  0.5014 -0.2172 -0.2637 -0.7952 -0.2354  0.2604  0.7127
#> 
#> Columns 19 to 20 -0.8295 -0.8099
#>  -0.2337  0.4406
#>  -0.7707  0.2328
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0722  0.4559  0.3166  0.7649 -0.5511 -0.3415 -0.6810 -0.7144  0.4775
#>   0.0978 -0.2892  0.6014 -0.1651 -0.6319 -0.5995  0.1310 -0.6320  0.1381
#>  -0.2787  0.3653 -0.6048 -0.2550 -0.3068 -0.3188  0.4365 -0.0545 -0.1988
#> 
#> Columns 10 to 18 -0.5000  0.6969  0.6251  0.6434  0.2352  0.4734  0.1890  0.0612  0.1187
#>   0.1631  0.4842  0.3248  0.2518  0.5183  0.1371  0.2105  0.2196 -0.3262
#>  -0.3813  0.2913 -0.2117  0.2004  0.3648 -0.6047 -0.0722 -0.0150  0.2552
#> 
#> Columns 19 to 20 -0.3983 -0.2409
#>  -0.3993 -0.4017
#>   0.5613  0.2221
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1665  0.3721  0.6359  0.2426 -0.3004 -0.1499  0.0923 -0.2783  0.4832
#>  -0.0051  0.2861 -0.0126 -0.2364 -0.1432 -0.2950  0.1762 -0.0526  0.4083
#>   0.1242 -0.0480 -0.3815  0.0075 -0.5554 -0.4163  0.4130  0.1565  0.2170
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1658 -0.1344  0.2463  0.0760  0.2888 -0.4817  0.0595 -0.3879  0.2239
#>   0.4211 -0.5440 -0.6111  0.1207  0.6019  0.3947 -0.5801  0.1673  0.0731
#>   0.2735  0.6450  0.3556  0.7939 -0.4682  0.2221 -0.4684  0.5688 -0.0090
#> 
#> Columns 10 to 18 -0.0626 -0.1103  0.0528  0.1181 -0.3039 -0.4247 -0.0848 -0.2331 -0.2813
#>  -0.5143 -0.5383 -0.0011 -0.3064 -0.4371 -0.5704 -0.0440  0.6039 -0.2204
#>   0.1545 -0.1634  0.4182  0.3028 -0.2091 -0.6934  0.3040 -0.1619 -0.5544
#> 
#> Columns 19 to 20 -0.4583  0.2871
#>  -0.5350 -0.3449
#>  -0.4698  0.4579
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2906  0.2411  0.0822 -0.0020 -0.2825 -0.1703  0.1655 -0.0418  0.3589
#>  -0.2055  0.3030  0.3095  0.0408 -0.6712 -0.0554  0.0656 -0.3380  0.5314
#>   0.0078  0.5386  0.2437  0.4013 -0.6295 -0.2060 -0.3206  0.0672  0.6075
#> 
#> Columns 10 to 18 -0.3241  0.3733 -0.1941  0.3711  0.4086 -0.4533  0.0448  0.0719  0.1214
#>  -0.5574 -0.0186 -0.2725  0.2956  0.3318 -0.5637  0.1084 -0.2533 -0.1798
#>  -0.1147  0.2918 -0.3157  0.6086 -0.3258 -0.4929  0.4891  0.0794 -0.0931
#> 
#> Columns 19 to 20 -0.2790 -0.1302
#>  -0.0462 -0.0665
#>  -0.4766 -0.0876
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>