RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0924  0.3643 -0.2927  0.0496 -0.5517  0.6239 -0.6317  0.3378 -0.8322
#>   0.2679 -0.5827  0.3396 -0.6547 -0.8326  0.2104 -0.4523  0.2977  0.2484
#>   0.4355 -0.9002  0.2436  0.6891 -0.0662  0.3890 -0.0678 -0.1811 -0.1059
#> 
#> Columns 10 to 18  0.7783 -0.5669  0.5228 -0.8857  0.4205  0.4666  0.2003 -0.3222  0.0740
#>  -0.1766 -0.6836  0.2121  0.3903 -0.5894 -0.6601 -0.0459 -0.5790 -0.7137
#>  -0.4796 -0.6607  0.2063  0.6002 -0.1472  0.0737 -0.3405 -0.8463  0.2219
#> 
#> Columns 19 to 20  0.0452 -0.2517
#>   0.5138 -0.4196
#>   0.1378  0.0272
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6812  0.2714 -0.2924  0.0966 -0.3430  0.4722 -0.2177  0.2804  0.4162
#>  -0.2178  0.3742 -0.0270 -0.5409 -0.1315  0.1626 -0.1028 -0.1306  0.4796
#>  -0.3386  0.0974  0.1944  0.2417 -0.0455  0.2720  0.4304  0.0874  0.7430
#> 
#> Columns 10 to 18 -0.1859 -0.2805  0.4505  0.0066  0.1577  0.5905 -0.0739 -0.4118 -0.6040
#>  -0.0662 -0.2917  0.0178 -0.3899  0.3012  0.1254  0.2861 -0.4987 -0.3591
#>  -0.5690 -0.1603  0.0627 -0.3922  0.3418  0.1889  0.3738 -0.0376 -0.1377
#> 
#> Columns 19 to 20 -0.1405  0.2419
#>   0.3130  0.0125
#>   0.0761 -0.3752
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1200 -0.5974  0.1440  0.1613 -0.1789 -0.1740  0.2961 -0.2986  0.1779
#>   0.3824 -0.7753  0.3043  0.3568 -0.3793  0.2887  0.2738  0.6027  0.2625
#>   0.0552 -0.5370  0.0151  0.4853 -0.2965 -0.1381  0.1337  0.1756  0.4117
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0882  0.0343  0.3263 -0.5259  0.2110 -0.5906 -0.4928 -0.7609  0.2162
#>   0.5766 -0.0421 -0.1251 -0.6032  0.2158 -0.0297 -0.2903  0.3117 -0.1932
#>   0.1174  0.0244  0.0949  0.2236 -0.0396 -0.3831 -0.1027  0.4049  0.2406
#> 
#> Columns 10 to 18  0.0939 -0.2539  0.0574  0.0510 -0.3447 -0.5954  0.4104 -0.3668  0.5205
#>  -0.2061  0.0997 -0.0422 -0.1614 -0.7451 -0.0561 -0.0511  0.3455  0.0761
#>   0.3397  0.2397  0.1253 -0.2482 -0.5343  0.0871  0.0423  0.3144 -0.3693
#> 
#> Columns 19 to 20  0.6885 -0.2002
#>   0.1893 -0.1140
#>   0.0666  0.2624
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3164 -0.4503 -0.0273  0.3285 -0.3123 -0.0101  0.1919  0.3528  0.4119
#>  -0.1899 -0.2463 -0.3187 -0.2034 -0.4846  0.2668  0.3126  0.1460  0.3686
#>  -0.1770 -0.5190  0.3326 -0.4871 -0.7102  0.3210  0.3705  0.2649  0.0471
#> 
#> Columns 10 to 18 -0.0516 -0.4167 -0.2292 -0.1149  0.1658  0.0404 -0.5072  0.1327 -0.1337
#>   0.2141 -0.1441 -0.3887  0.0990  0.4902 -0.0390 -0.2064 -0.5100 -0.4849
#>   0.2875 -0.2733 -0.2344 -0.1836  0.3747 -0.3100 -0.0341 -0.4158 -0.2542
#> 
#> Columns 19 to 20  0.3702  0.2810
#>  -0.1750 -0.1422
#>   0.0745  0.0161
#> [ CPUFloatType{2,3,20} ]
#>