RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5408 -0.8382 -0.3799  0.7562 -0.5784 -0.2574 -0.7102 -0.2183 -0.3388
#>   0.0659  0.5090  0.6206  0.4716  0.6869 -0.2349 -0.3894  0.7593  0.4825
#>   0.5767 -0.0178 -0.5918 -0.1921  0.8141 -0.2007 -0.8103 -0.1975  0.3299
#> 
#> Columns 10 to 18  0.5153  0.0818 -0.0580  0.9116  0.1763 -0.2822  0.7288 -0.2426  0.5348
#>  -0.6483  0.2518  0.2011  0.4942  0.5445 -0.4975  0.2031  0.4350  0.0679
#>   0.8901  0.3042 -0.2410  0.6286  0.0321 -0.2847  0.5090  0.5180 -0.8054
#> 
#> Columns 19 to 20  0.7046  0.6873
#>   0.4513  0.4084
#>  -0.8255 -0.2492
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5887  0.0752  0.4305  0.0339 -0.1681 -0.4518  0.1612  0.5109 -0.0439
#>  -0.2954 -0.4048 -0.5054 -0.3635 -0.0655 -0.0454 -0.1245  0.3406 -0.2208
#>   0.7185  0.2176 -0.4013  0.3223  0.3789 -0.0107 -0.0201 -0.3245 -0.6535
#> 
#> Columns 10 to 18 -0.4702 -0.0150  0.4221  0.1485  0.2226  0.0263  0.3851 -0.1044  0.0757
#>   0.2208  0.6137 -0.3028  0.3943 -0.1038 -0.4190  0.5248  0.0349  0.3039
#>   0.0250  0.3098  0.5794  0.1634  0.2046  0.4411  0.0424 -0.2502  0.0783
#> 
#> Columns 19 to 20  0.3862 -0.2959
#>  -0.4976  0.2141
#>  -0.3912 -0.3364
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2130 -0.5587  0.0645 -0.3228 -0.0492  0.0989 -0.0910  0.4720 -0.3427
#>   0.4516  0.0780  0.2331  0.0382  0.0969 -0.2463 -0.3389 -0.2809 -0.2594
#>   0.2458  0.3446  0.5220 -0.4227  0.5598 -0.0794  0.4179  0.1342 -0.0125
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6568  0.4438  0.8095  0.3745  0.7443  0.1269 -0.7733 -0.4623  0.3981
#>   0.3096 -0.4989 -0.0679 -0.5739 -0.1206  0.7799 -0.0035  0.3477 -0.0614
#>  -0.0943 -0.7251 -0.0536 -0.1360  0.0024  0.3525 -0.0383  0.2531 -0.2965
#> 
#> Columns 10 to 18 -0.4761 -0.1888 -0.2713 -0.4234 -0.0496  0.6341 -0.5135 -0.6603  0.3222
#>   0.0556 -0.1704 -0.0287  0.0757 -0.6905 -0.0091 -0.8066 -0.6336  0.3587
#>   0.1950 -0.2219 -0.0030  0.0650 -0.4423 -0.2127 -0.1756 -0.3108  0.0012
#> 
#> Columns 19 to 20 -0.5867  0.4769
#>   0.4760  0.0752
#>  -0.0741  0.1314
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1683  0.1656  0.3848 -0.3963  0.5047 -0.2446 -0.1718  0.5587  0.2139
#>  -0.2132 -0.1584 -0.1589  0.2785  0.3083  0.1021  0.0341  0.2067 -0.2791
#>  -0.3057 -0.1109 -0.2051  0.0109 -0.1515  0.1645 -0.1044  0.0693 -0.0575
#> 
#> Columns 10 to 18  0.1637 -0.2129  0.1733 -0.1711  0.1918 -0.4967  0.5221  0.2615 -0.0615
#>   0.3605  0.0652  0.2556 -0.4963  0.4053 -0.0535  0.1134 -0.0701 -0.0487
#>   0.3020  0.1462 -0.0406 -0.3446 -0.1048 -0.1943 -0.0982  0.2829  0.1735
#> 
#> Columns 19 to 20 -0.2139  0.2079
#>  -0.2585 -0.1208
#>  -0.1895 -0.2320
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>