RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6760 -0.5367  0.4839 -0.4601  0.2121  0.0609  0.1824 -0.0343  0.4126
#>   0.2277 -0.3111 -0.1826  0.3426 -0.4805  0.0802  0.2553  0.0448  0.0990
#>  -0.4637  0.0949  0.3147  0.8717  0.9937 -0.8357 -0.8791  0.8711  0.5305
#> 
#> Columns 10 to 18 -0.2624 -0.1001  0.5080  0.1110 -0.3310 -0.1226  0.3271  0.3657 -0.3655
#>   0.7809 -0.0539 -0.4903 -0.7761 -0.7455 -0.4340  0.8556 -0.7383 -0.3865
#>   0.3938  0.3415 -0.2974 -0.6840  0.8267  0.2347 -0.6424  0.8028 -0.6802
#> 
#> Columns 19 to 20  0.4812  0.5099
#>  -0.4141  0.7251
#>   0.6546  0.8405
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4746  0.0075  0.2003  0.5297  0.0442  0.1719  0.0864 -0.0295 -0.1313
#>   0.0005  0.0931  0.7296  0.2881  0.0718  0.1957 -0.0233  0.2067  0.2058
#>   0.2862 -0.1427  0.3449  0.5555 -0.4511  0.5142  0.1057 -0.5950 -0.0694
#> 
#> Columns 10 to 18  0.2838  0.2551 -0.2744  0.0960 -0.2235 -0.4189 -0.0451  0.0178 -0.5399
#>   0.4491 -0.1394  0.1473  0.1364 -0.4445  0.2409 -0.5454 -0.1283 -0.5437
#>  -0.0628  0.5234 -0.6421 -0.0693 -0.3464  0.3885 -0.4337 -0.5727 -0.0381
#> 
#> Columns 19 to 20  0.2661 -0.0882
#>   0.2698 -0.3005
#>   0.1518  0.3299
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2529 -0.2414  0.6935  0.1485  0.6255 -0.1982 -0.1559  0.2412 -0.0917
#>   0.0830  0.1978  0.1253  0.4425  0.2885 -0.1741 -0.0690 -0.1275  0.0228
#>   0.0767 -0.1992  0.3077  0.4062  0.2752  0.0671 -0.0140 -0.3461 -0.0946
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2334  0.3097 -0.0466 -0.0550  0.4919 -0.8294 -0.4684  0.2973  0.1248
#>   0.2875  0.2853 -0.4464 -0.0407  0.2841 -0.3810  0.1521  0.3931  0.2500
#>  -0.2986 -0.7566 -0.2068 -0.1213  0.0176 -0.4575 -0.4127 -0.0663  0.1409
#> 
#> Columns 10 to 18  0.0211 -0.5297  0.0083  0.7790 -0.5491 -0.1988  0.5626  0.0812  0.3068
#>   0.6257 -0.5639  0.1986  0.8404 -0.0800  0.0553  0.0376 -0.1685 -0.0871
#>   0.0615  0.3757 -0.5782 -0.2348 -0.6676 -0.2914  0.1680 -0.1947  0.0119
#> 
#> Columns 19 to 20  0.5765  0.0130
#>  -0.1712 -0.1757
#>   0.3469  0.2385
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0859  0.1833  0.1224  0.0221  0.2374  0.0544 -0.0198  0.1570  0.1096
#>   0.2068  0.0428  0.0883  0.1205  0.2228 -0.1563 -0.3062  0.2728  0.3935
#>  -0.2783 -0.0848  0.4385  0.1211  0.5378 -0.1505  0.5322 -0.0284 -0.1317
#> 
#> Columns 10 to 18  0.1390 -0.2422 -0.2626 -0.7022 -0.2679  0.1257 -0.1248 -0.3444 -0.2864
#>   0.1846  0.0972 -0.2234 -0.5137 -0.3587 -0.2092 -0.4072 -0.3152 -0.2216
#>  -0.1591 -0.0783 -0.4387 -0.0770 -0.1553  0.6543 -0.2042 -0.3201 -0.2530
#> 
#> Columns 19 to 20  0.1939  0.1738
#>   0.3695  0.0862
#>  -0.3646 -0.0274
#> [ CPUFloatType{2,3,20} ]
#>