RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0644 -0.8708  0.4467 -0.8905 -0.0284  0.2211 -0.2545 -0.3488  0.7430
#>  -0.3091 -0.8042  0.5062  0.1456 -0.0066 -0.5712 -0.7654  0.1648 -0.2713
#>  -0.7217  0.8578 -0.2210  0.6238  0.3509 -0.6624  0.3463 -0.5968 -0.0626
#> 
#> Columns 10 to 18  0.0928  0.6033  0.5093  0.7385  0.7050 -0.0572 -0.1522 -0.6260 -0.8152
#>  -0.7435  0.2191 -0.2177  0.4506 -0.1089 -0.2681 -0.1858 -0.2120  0.2733
#>   0.1778 -0.4515  0.6614 -0.3463  0.1194  0.5352 -0.6161  0.4320  0.2864
#> 
#> Columns 19 to 20 -0.4916  0.2849
#>  -0.6126  0.7642
#>   0.7090  0.3519
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3985 -0.0623  0.5274  0.7757  0.2253  0.0109  0.4042 -0.1601 -0.1005
#>   0.7863 -0.0662  0.7260  0.4621  0.0391  0.5321 -0.1672 -0.6295  0.3654
#>  -0.1458 -0.4385  0.2083  0.0530  0.4118  0.4361  0.0535 -0.2253  0.7050
#> 
#> Columns 10 to 18 -0.3010 -0.0262 -0.2434 -0.2887 -0.7331 -0.2104 -0.5839 -0.1378  0.6244
#>  -0.0143 -0.0345 -0.8574 -0.1895 -0.6424 -0.1487  0.3978 -0.0062 -0.4129
#>  -0.1552  0.5257 -0.3569 -0.2348 -0.1521 -0.0576 -0.4330 -0.0845 -0.4514
#> 
#> Columns 19 to 20  0.2376 -0.0333
#>   0.7000  0.1201
#>   0.6212 -0.3872
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3688 -0.3710 -0.1281 -0.1334 -0.0439  0.3552  0.2755 -0.5343  0.1760
#>   0.7152 -0.8014 -0.2995 -0.2028 -0.0978  0.4480 -0.4662 -0.1323  0.6949
#>   0.0415 -0.2934 -0.1469  0.1766  0.2590  0.4515  0.4266 -0.2369  0.1913
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8093 -0.1429  0.5126 -0.7409  0.3218 -0.1824  0.7133 -0.3436  0.1005
#>  -0.1479  0.3177  0.1377  0.6448 -0.6256 -0.0617  0.2082  0.3498 -0.0741
#>   0.1828  0.3358  0.0842  0.4275 -0.3800  0.0320 -0.6723  0.4758 -0.2404
#> 
#> Columns 10 to 18  0.6289 -0.5449  0.8253  0.1404 -0.0111  0.6014 -0.0427 -0.1037 -0.5980
#>   0.2858  0.3546 -0.2940  0.5212  0.2289 -0.1451  0.4895 -0.1279  0.0151
#>  -0.2775 -0.3190  0.1559  0.5463  0.0576 -0.1720 -0.1306 -0.0628  0.0169
#> 
#> Columns 19 to 20 -0.6851 -0.5832
#>   0.1653  0.5088
#>  -0.0071  0.0060
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5351 -0.6654  0.2495  0.1190 -0.3301  0.3729 -0.2634 -0.5692  0.7525
#>   0.1022  0.1343  0.0302  0.1629 -0.2037 -0.3907  0.0518 -0.2511 -0.3562
#>   0.0611 -0.3903 -0.2213  0.1747  0.0586  0.1278 -0.1001 -0.2011 -0.0478
#> 
#> Columns 10 to 18 -0.1641 -0.1085 -0.6207  0.3128 -0.3796  0.2496 -0.0622  0.1592 -0.3546
#>   0.5170  0.0499  0.1842  0.0747  0.2933 -0.4492 -0.0628 -0.0696 -0.3662
#>  -0.2604 -0.0910  0.2671  0.3637  0.2284 -0.3606 -0.5264 -0.2065  0.0563
#> 
#> Columns 19 to 20  0.3544  0.4611
#>   0.1016 -0.1494
#>  -0.3040 -0.3922
#> [ CPUFloatType{2,3,20} ]
#>