RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8632  0.3915 -0.1843  0.3043  0.7445  0.4889  0.7855 -0.2183 -0.6717
#>   0.5961  0.3596 -0.3534 -0.6031  0.1287 -0.2378  0.8612  0.3927 -0.2000
#>   0.1290 -0.2816  0.9479 -0.4394 -0.7786  0.7990  0.5219  0.8691  0.4200
#> 
#> Columns 10 to 18  0.5067 -0.8151 -0.6813  0.3916 -0.4814 -0.6896 -0.1054 -0.9225  0.3987
#>  -0.3986  0.5969 -0.1295 -0.4747 -0.6868  0.6671  0.9608  0.4509 -0.8185
#>   0.2374  0.3972 -0.4167  0.2703  0.6027 -0.0227 -0.2131  0.4866 -0.3522
#> 
#> Columns 19 to 20  0.2309  0.2988
#>  -0.2302  0.1902
#>   0.1540  0.3401
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1938  0.5071 -0.1384  0.1387  0.1695 -0.0776  0.4758 -0.0156 -0.5408
#>   0.0381 -0.0880 -0.0833 -0.5086  0.1871  0.1594 -0.1546  0.7601 -0.3696
#>   0.6895  0.5989  0.3813  0.6193 -0.2575  0.3903  0.2980  0.6763  0.3721
#> 
#> Columns 10 to 18 -0.3101 -0.4844 -0.5209  0.1982 -0.2307 -0.5724  0.2298 -0.6644  0.1731
#>  -0.0748  0.0269 -0.0733 -0.3398 -0.2354  0.4837  0.7447 -0.4934 -0.6305
#>  -0.0047 -0.1367 -0.1804  0.2313  0.3449 -0.4094 -0.1627 -0.4431 -0.0665
#> 
#> Columns 19 to 20  0.4487  0.4741
#>   0.0522 -0.4587
#>   0.2663 -0.0463
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3701  0.0173  0.2344  0.1476  0.1303  0.3092  0.4669 -0.0440 -0.5744
#>   0.3021  0.0279  0.0744 -0.1874 -0.0454  0.1973 -0.0938  0.6960 -0.1001
#>   0.4824  0.5584  0.3160  0.4132 -0.2624  0.4666  0.3269  0.6231  0.0304
#> 
#> Columns 10 to 18 -0.1895 -0.4840 -0.4338  0.3951 -0.0370 -0.1106  0.2804 -0.7316 -0.1943
#>  -0.4826 -0.3274 -0.4963  0.0495  0.0499 -0.0797  0.5866 -0.5700 -0.6528
#>  -0.2509 -0.0647 -0.2250  0.2422  0.3436 -0.0558 -0.0750 -0.5384  0.2148
#> 
#> Columns 19 to 20  0.4306 -0.3075
#>   0.0100 -0.0329
#>   0.6925  0.2668
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.0817  0.0087  0.4342  0.3156 -0.2054  0.4280  0.2253  0.4189 -0.1605
#>   0.2420 -0.0618  0.2840  0.0863 -0.1413  0.6450  0.1243  0.7762 -0.1055
#>   0.1946  0.4120  0.4445  0.2163 -0.0745  0.3722 -0.2051  0.6255  0.1528
#> 
#> Columns 10 to 18 -0.5755 -0.4312 -0.3656  0.0357  0.2886 -0.3300 -0.0302 -0.0466 -0.0640
#>  -0.4272 -0.4477 -0.5410  0.4288  0.5747 -0.2675  0.4738 -0.5490 -0.4452
#>  -0.5083 -0.2301 -0.3301  0.0988  0.2058 -0.2807 -0.0725 -0.6231  0.0598
#> 
#> Columns 19 to 20  0.5861  0.2883
#>   0.0749 -0.1205
#>   0.7839 -0.0752
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.4633  0.0148  0.2988  0.1738 -0.0171  0.3857  0.4436 -0.1052 -0.3697
#>   0.4388  0.1635  0.2645  0.6325 -0.1622  0.5991  0.5669  0.3383 -0.1884
#>   0.3283 -0.0441  0.3801  0.4120  0.2052  0.2883 -0.2517  0.1785 -0.1536
#> 
#> Columns 10 to 18 -0.0993 -0.3836 -0.3051 -0.1341 -0.0937 -0.3693 -0.3204 -0.4197 -0.1949
#>  -0.2169 -0.5127 -0.5784  0.6481  0.4354 -0.2607  0.0910 -0.5296 -0.1353
#>  -0.4500 -0.2745 -0.4688  0.0599  0.3477 -0.1991 -0.2825 -0.6111  0.3215
#> 
#> Columns 19 to 20  0.3467 -0.6367
#>   0.1238  0.0829
#>   0.6310 -0.3190
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7609 -0.2838 -0.0239  0.0645  0.0793  0.6334 -0.2708  0.5658 -0.4879
#>  -0.3875  0.3600 -0.7532  0.7044  0.4486  0.2431  0.1267 -0.6482  0.3584
#>  -0.2083 -0.5876  0.1607  0.2945  0.7873  0.1900 -0.3695  0.3025  0.5961
#> 
#> Columns 10 to 18  0.7119 -0.3135  0.0118  0.5489 -0.3813 -0.4337  0.0287  0.3841 -0.2247
#>   0.6930  0.0972 -0.8746  0.1846 -0.8284 -0.2289 -0.6612 -0.3000 -0.6738
#>   0.5774 -0.3245  0.4584 -0.1863  0.1446  0.1564  0.3423 -0.7717  0.0546
#> 
#> Columns 19 to 20 -0.4619  0.2965
#>  -0.4088  0.4089
#>   0.2912  0.4310
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4633  0.0148  0.2988  0.1738 -0.0171  0.3857  0.4436 -0.1052 -0.3697
#>   0.4388  0.1635  0.2645  0.6325 -0.1622  0.5991  0.5669  0.3383 -0.1884
#>   0.3283 -0.0441  0.3801  0.4120  0.2052  0.2883 -0.2517  0.1785 -0.1536
#> 
#> Columns 10 to 18 -0.0993 -0.3836 -0.3051 -0.1341 -0.0937 -0.3693 -0.3204 -0.4197 -0.1949
#>  -0.2169 -0.5127 -0.5784  0.6481  0.4354 -0.2607  0.0910 -0.5296 -0.1353
#>  -0.4500 -0.2745 -0.4688  0.0599  0.3477 -0.1991 -0.2825 -0.6111  0.3215
#> 
#> Columns 19 to 20  0.3467 -0.6367
#>   0.1238  0.0829
#>   0.6310 -0.3190
#> [ CPUFloatType{2,3,20} ]
#>