RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5240 -0.1222  0.7113  0.3355 -0.2601 -0.8521 -0.4104  0.8557  0.6391
#>   0.3906 -0.6609 -0.7145 -0.2122  0.3211 -0.8035  0.5892  0.1622  0.7019
#>   0.8382 -0.9192  0.1262 -0.4827 -0.7819  0.0132  0.6487  0.7554 -0.2927
#> 
#> Columns 10 to 18 -0.2927 -0.0029  0.4334  0.6671 -0.8762 -0.0724 -0.8011 -0.0511 -0.2746
#>  -0.9149  0.0026 -0.5843 -0.4313  0.8480  0.5761 -0.2266 -0.8072 -0.4485
#>  -0.4293  0.7807 -0.7356  0.3421  0.1501  0.4902  0.6315  0.2469 -0.0564
#> 
#> Columns 19 to 20  0.3456  0.2476
#>  -0.1318  0.1769
#>  -0.1799 -0.5119
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4600 -0.1669  0.2091 -0.4578 -0.6172  0.2080 -0.1805  0.6314  0.4550
#>   0.4628 -0.5406 -0.0261 -0.4957 -0.3015  0.1401 -0.4368  0.0099  0.6820
#>   0.1674 -0.7196  0.0736  0.4625  0.1203  0.3588 -0.0476  0.5635  0.1124
#> 
#> Columns 10 to 18  0.1267  0.2992  0.3518  0.0782 -0.5038  0.3972 -0.0212  0.1900 -0.0553
#>  -0.0302  0.4596 -0.7717 -0.3231 -0.5959 -0.1916 -0.2779  0.1596 -0.1478
#>  -0.5501  0.1026 -0.4478 -0.6078 -0.6270  0.7940 -0.5681 -0.1195  0.5171
#> 
#> Columns 19 to 20  0.2041 -0.4443
#>  -0.2574  0.3329
#>  -0.3884 -0.5085
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0106 -0.1525 -0.4407  0.3907  0.0876 -0.2528 -0.2763  0.1722  0.5045
#>  -0.2169 -0.8217 -0.3587 -0.6511  0.1763 -0.0623 -0.5550  0.5177  0.6575
#>   0.6991 -0.1470 -0.1562 -0.2454 -0.1062 -0.7349 -0.0886  0.5879  0.5103
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2078  0.2185 -0.1613 -0.2715  0.1440  0.8795  0.1129 -0.8332  0.7505
#>   0.5869  0.3663 -0.1776 -0.0413  0.6003  0.4431  0.4220 -0.8974 -0.0346
#>  -0.0252  0.3032  0.0559 -0.4541  0.2964  0.3703  0.4547 -0.5691  0.0130
#> 
#> Columns 10 to 18 -0.2422  0.1266  0.6037 -0.0479  0.6217  0.6268  0.0004  0.2451 -0.6597
#>  -0.7093  0.1363 -0.7138  0.8566  0.7445  0.3776  0.2402  0.1349  0.2268
#>   0.1453  0.1645 -0.2514  0.5741  0.0300  0.5501 -0.1736  0.0359  0.3357
#> 
#> Columns 19 to 20  0.5773 -0.0341
#>   0.7441  0.3134
#>  -0.0271  0.5759
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2672 -0.2153 -0.1763  0.6020  0.0417  0.3358 -0.1722  0.0572  0.3950
#>   0.2254 -0.2293  0.0129  0.4519 -0.1928 -0.2487 -0.4817  0.1123 -0.0178
#>   0.1619 -0.0880 -0.0304  0.4824 -0.1793 -0.3242 -0.3410  0.5013  0.2952
#> 
#> Columns 10 to 18 -0.5303 -0.1626 -0.0937 -0.1777 -0.3785 -0.0082 -0.0677  0.4110  0.3591
#>  -0.3402 -0.1994 -0.4921 -0.0317 -0.4999 -0.3491  0.2325 -0.3261  0.5368
#>  -0.3233 -0.0735  0.0964 -0.1809 -0.7129 -0.1970 -0.3129  0.1272  0.3501
#> 
#> Columns 19 to 20  0.2918 -0.3673
#>   0.3466 -0.2244
#>   0.0393 -0.4381
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>