RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0349  0.2449 -0.1930 -0.3415 -0.5714  0.7511 -0.4145  0.5218  0.3212
#>   0.2030  0.5050  0.3829  0.2667  0.2030  0.2078 -0.0763  0.4492 -0.3061
#>  -0.2678  0.4950  0.6954  0.1008  0.0079 -0.7498 -0.5578 -0.0128  0.4404
#> 
#> Columns 10 to 18 -0.8087 -0.6406 -0.9389 -0.6295  0.2326  0.6953 -0.3096  0.0326 -0.1057
#>  -0.7265 -0.1091 -0.4328 -0.2791 -0.7921 -0.0545  0.3218 -0.4124 -0.4338
#>  -0.0849 -0.2589  0.5183 -0.8516 -0.7327 -0.2676 -0.3735 -0.5953  0.7276
#> 
#> Columns 19 to 20 -0.9807  0.0868
#>  -0.1315  0.6923
#>  -0.5239  0.3313
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.8339 -0.6265  0.8008  0.2454  0.5592  0.2068 -0.5123 -0.5775  0.3787
#>   0.6387  0.3082  0.8701 -0.0007 -0.2553 -0.0597 -0.4670 -0.7206 -0.4620
#>  -0.2723 -0.1501  0.4820  0.1243 -0.2669  0.2765 -0.2869 -0.7301 -0.6075
#> 
#> Columns 10 to 18 -0.6766  0.4835 -0.0464  0.1799 -0.2297 -0.5352  0.4702 -0.7613 -0.2966
#>   0.0858  0.0176  0.3153 -0.1011 -0.5261  0.4614  0.1546 -0.8822 -0.7953
#>   0.4485  0.4335 -0.0158 -0.0132  0.0831  0.1227  0.1914 -0.8464 -0.5144
#> 
#> Columns 19 to 20 -0.6256  0.4531
#>  -0.0327  0.5812
#>   0.1426  0.4689
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1972 -0.1768  0.3962 -0.0477  0.3484 -0.1884 -0.4264 -0.4815 -0.2106
#>   0.5833 -0.3313 -0.1681 -0.1416 -0.2261 -0.1411 -0.4901 -0.5403 -0.2855
#>   0.6291 -0.4331 -0.2051 -0.0987  0.1211 -0.0788 -0.3492 -0.5389 -0.3223
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4564  0.5953  0.5880 -0.9038  0.4714 -0.0986 -0.0972  0.7025 -0.4181
#>   0.3188  0.1847 -0.1080  0.1251 -0.1663 -0.7074  0.3443 -0.2403 -0.1124
#>   0.4465  0.5038  0.3656 -0.2426  0.4855 -0.5957 -0.2791 -0.3847  0.1817
#> 
#> Columns 10 to 18 -0.8686 -0.7642  0.5776 -0.4006  0.8566 -0.7586  0.3651  0.0754 -0.0948
#>   0.2687 -0.6683 -0.1140  0.5806  0.0231 -0.1265  0.0092  0.0360  0.7111
#>  -0.2091 -0.6656  0.3801 -0.5058 -0.3484 -0.7274 -0.7216 -0.4640  0.2882
#> 
#> Columns 19 to 20  0.7252  0.5855
#>   0.6170 -0.4159
#>   0.8326  0.0550
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5101  0.2764  0.7125 -0.1132 -0.0373 -0.0961 -0.8497 -0.1617 -0.2726
#>   0.0185 -0.3425  0.1455  0.0304  0.3765 -0.1260 -0.5288 -0.2390 -0.3038
#>   0.1793  0.2014  0.6625 -0.2740  0.0416 -0.2089 -0.5569 -0.1245 -0.2897
#> 
#> Columns 10 to 18 -0.1360 -0.3762 -0.0726 -0.4955 -0.2191  0.6146  0.0349 -0.7246 -0.6326
#>  -0.2054  0.1131 -0.3256 -0.3745  0.1904 -0.1473  0.6285 -0.0964 -0.1818
#>  -0.2834 -0.2964 -0.6428 -0.5747 -0.2849  0.1530  0.5707 -0.5280 -0.1746
#> 
#> Columns 19 to 20 -0.1972  0.4975
#>  -0.1680  0.3542
#>  -0.5718  0.4176
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>