RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0548  0.3017 -0.9517 -0.7818 -0.6874  0.4712 -0.7881  0.3979 -0.7836
#>  -0.5948  0.1516  0.6240 -0.7104 -0.4476  0.2109  0.2104 -0.6182 -0.0597
#>  -0.0717  0.1354 -0.6815  0.7721  0.7676  0.9255  0.3777 -0.0855  0.3740
#> 
#> Columns 10 to 18  0.3728 -0.2969 -0.2590  0.0899 -0.1555  0.7143  0.3609  0.2488 -0.9317
#>  -0.9130 -0.9190  0.7464  0.6061  0.8770 -0.2886 -0.9685 -0.1210 -0.5910
#>   0.7216 -0.0794  0.3015  0.6902  0.3527 -0.4678 -0.1178 -0.3437  0.0338
#> 
#> Columns 19 to 20  0.9283  0.7032
#>   0.7606 -0.6421
#>   0.4256 -0.1473
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3857  0.3521  0.1872  0.0886 -0.3244  0.6647  0.5449 -0.0075 -0.3876
#>  -0.1096 -0.4543  0.2941 -0.2369 -0.6863  0.7907  0.6329  0.2203 -0.4950
#>  -0.2989 -0.5923 -0.1376 -0.1063 -0.6700  0.3237 -0.2445 -0.0561  0.1684
#> 
#> Columns 10 to 18 -0.1015  0.5710 -0.4154 -0.0068 -0.2293  0.5426  0.3977  0.6859  0.2315
#>  -0.3623  0.2440 -0.1656 -0.3647  0.0829  0.3860  0.4324  0.6775 -0.3429
#>  -0.6123  0.2691 -0.6484 -0.3063 -0.5223 -0.0003  0.1127  0.2127  0.2282
#> 
#> Columns 19 to 20 -0.2486  0.0082
#>  -0.2900  0.3811
#>   0.6921  0.2359
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4583 -0.2231  0.1759 -0.2217  0.0441  0.3069 -0.1163 -0.3407 -0.4134
#>  -0.4666 -0.2828  0.3053  0.2860 -0.2527  0.4223  0.1282 -0.0715 -0.2371
#>  -0.2701 -0.0343  0.4650  0.2564 -0.0202  0.0894  0.4887 -0.2631  0.0755
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2790 -0.1761  0.2464  0.4232  0.4259  0.0725 -0.1823  0.3391  0.3914
#>  -0.0569 -0.2119  0.0473  0.2320  0.1596 -0.7603 -0.4470  0.6353 -0.1035
#>   0.2488 -0.6912 -0.2931 -0.5636  0.0428 -0.5199 -0.5797  0.3314 -0.7634
#> 
#> Columns 10 to 18  0.7421 -0.1857  0.0966  0.4375 -0.7878 -0.2558 -0.1958  0.1314 -0.4982
#>   0.6879 -0.2465 -0.1903 -0.1128 -0.6939 -0.5438  0.4357  0.6866 -0.2626
#>  -0.0597  0.0230 -0.1364  0.7820 -0.5527 -0.3002  0.0223 -0.1371  0.1241
#> 
#> Columns 19 to 20 -0.6773 -0.1580
#>  -0.8113  0.8919
#>  -0.2063  0.2365
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3967  0.0674 -0.0466 -0.2087 -0.3974  0.7835  0.0633  0.4376 -0.2381
#>  -0.4417 -0.0238 -0.5392  0.1202 -0.3074  0.4751 -0.2283 -0.0073  0.1335
#>  -0.6192 -0.1370  0.1292 -0.0831  0.3681  0.4637 -0.1700 -0.3836  0.0277
#> 
#> Columns 10 to 18 -0.1206  0.0541  0.2622 -0.0183  0.0705  0.3709  0.1327  0.3730  0.4087
#>  -0.1612  0.0381  0.5124  0.2639  0.1902 -0.1570 -0.2426 -0.1619 -0.0007
#>  -0.3734 -0.1976  0.3102  0.0109  0.1842  0.1179 -0.4937  0.0764  0.1376
#> 
#> Columns 19 to 20  0.1991 -0.5330
#>   0.4968  0.0431
#>   0.4355 -0.3496
#> [ CPUFloatType{2,3,20} ]
#>