RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4877 -0.1627 -0.0530  0.5755  0.0543 -0.0618  0.2896 -0.1261 -0.7778
#>   0.2501 -0.8641 -0.0456 -0.4657  0.5011 -0.3602  0.9503  0.5001 -0.3406
#>   0.5830  0.7052  0.0944  0.0980  0.0667 -0.7681 -0.1173  0.4567 -0.2057
#> 
#> Columns 10 to 18 -0.9300 -0.6089  0.7966  0.2307  0.2215  0.6324 -0.3141 -0.2183  0.7976
#>  -0.2125  0.6275 -0.5022  0.4615 -0.3377  0.6843  0.8910 -0.0104 -0.9599
#>   0.4918 -0.7418 -0.8776 -0.7558 -0.0101 -0.8784 -0.0353 -0.7162  0.1504
#> 
#> Columns 19 to 20 -0.0709 -0.1785
#>  -0.7128 -0.6373
#>   0.6243  0.1459
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1573 -0.1812  0.0202  0.6961  0.4059  0.3064  0.5039  0.0505 -0.5853
#>   0.2807  0.1899 -0.2023  0.1402 -0.0199 -0.0060 -0.1391  0.8354  0.2415
#>   0.4270 -0.0104  0.0678 -0.0773 -0.0888  0.1487  0.4094  0.4897 -0.0358
#> 
#> Columns 10 to 18  0.2566 -0.3047 -0.3271 -0.5501 -0.2942 -0.1819 -0.2429 -0.1733 -0.2125
#>  -0.5275 -0.3858  0.0461 -0.1418  0.3670 -0.2974  0.8435 -0.0957 -0.6366
#>  -0.6314 -0.2835 -0.2613  0.1993  0.2814 -0.1837  0.3243 -0.5067 -0.1278
#> 
#> Columns 19 to 20  0.4061  0.1949
#>  -0.1234 -0.0722
#>   0.6947 -0.3657
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3905  0.3504 -0.1212 -0.4368 -0.0693  0.2235  0.4240  0.3874 -0.2448
#>   0.1343  0.3250  0.2971  0.5700  0.1002  0.0640  0.5305  0.2505 -0.1645
#>   0.2175  0.0593  0.5876  0.2326 -0.1650  0.6851  0.1890  0.6397  0.2586
#> 
#> Columns 10 to 18 -0.5646 -0.2539  0.0487 -0.4829  0.0369 -0.4659 -0.0331 -0.1721 -0.1240
#>  -0.3021 -0.1605 -0.1170  0.4131 -0.1019  0.1785  0.1910 -0.2854 -0.1831
#>  -0.5094 -0.2151 -0.4593  0.0124  0.1632 -0.1287  0.3613 -0.5534 -0.3811
#> 
#> Columns 19 to 20  0.4611  0.2994
#>   0.4334 -0.4413
#>   0.4650  0.1126
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.6108  0.0674  0.3641  0.5603 -0.1232 -0.0154  0.7690  0.0388 -0.5142
#>   0.4280  0.5296  0.0895  0.2167 -0.1409  0.3144  0.1595  0.6280  0.1455
#>   0.5700  0.1245  0.5507  0.6404  0.0179  0.2646  0.5333  0.1159 -0.1363
#> 
#> Columns 10 to 18  0.1095  0.0813 -0.2670  0.2371 -0.4352  0.3180  0.1020 -0.4450 -0.2348
#>  -0.3008 -0.0194  0.0710 -0.3012  0.1791 -0.2316  0.3149 -0.2799 -0.1587
#>  -0.3999 -0.3409 -0.5413  0.5080  0.4839 -0.0308  0.0260 -0.3878  0.0581
#> 
#> Columns 19 to 20  0.6377 -0.0575
#>   0.7190  0.3862
#>   0.5313  0.0827
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.4471 -0.0167  0.1827  0.2594  0.1037  0.4192  0.1155  0.3939 -0.1547
#>   0.3086  0.3104  0.3495  0.1594  0.3132  0.5137 -0.0179  0.5054 -0.2664
#>  -0.0149  0.0572  0.3493  0.5181  0.0010  0.5235  0.2273  0.0301 -0.1175
#> 
#> Columns 10 to 18 -0.5588 -0.4458 -0.2150 -0.4510  0.4377 -0.0917 -0.0057  0.1272  0.0309
#>  -0.4474 -0.0633 -0.5297  0.2659  0.1231  0.2036  0.2641 -0.7694  0.1181
#>  -0.4893 -0.5990 -0.2562 -0.1667  0.2175 -0.3138 -0.0315  0.1428  0.0058
#> 
#> Columns 19 to 20  0.5366  0.4754
#>   0.7153  0.1051
#>   0.6400 -0.3832
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.0130  0.3503  0.1768  0.3579 -0.1907  0.1535 -0.0034 -0.3353  0.5145
#>   0.6483 -0.8074  0.0065  0.7384  0.5390 -0.0656  0.0725 -0.5277  0.4422
#>  -0.5912  0.0601 -0.3765  0.5111 -0.4954  0.7999 -0.5968 -0.3063 -0.4828
#> 
#> Columns 10 to 18 -0.1473 -0.1945 -0.0750 -0.4113 -0.1496 -0.2160  0.2403 -0.0216  0.0887
#>  -0.1223 -0.3322  0.1856  0.4030 -0.0582 -0.0730 -0.3305 -0.6625 -0.1000
#>   0.4813  0.2623  0.0420 -0.2888 -0.1284 -0.2785 -0.1803  0.3335  0.6188
#> 
#> Columns 19 to 20 -0.3995 -0.2458
#>   0.4327  0.3473
#>  -0.5389 -0.0948
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4471 -0.0167  0.1827  0.2594  0.1037  0.4192  0.1155  0.3939 -0.1547
#>   0.3086  0.3104  0.3495  0.1594  0.3132  0.5137 -0.0179  0.5054 -0.2664
#>  -0.0149  0.0572  0.3493  0.5181  0.0010  0.5235  0.2273  0.0301 -0.1175
#> 
#> Columns 10 to 18 -0.5588 -0.4458 -0.2150 -0.4510  0.4377 -0.0917 -0.0057  0.1272  0.0309
#>  -0.4474 -0.0633 -0.5297  0.2659  0.1231  0.2036  0.2641 -0.7694  0.1181
#>  -0.4893 -0.5990 -0.2562 -0.1667  0.2175 -0.3138 -0.0315  0.1428  0.0058
#> 
#> Columns 19 to 20  0.5366  0.4754
#>   0.7153  0.1051
#>   0.6400 -0.3832
#> [ CPUFloatType{2,3,20} ]
#>