RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4508 -0.1414 -0.4951  0.2936  0.6356 -0.5158 -0.2683 -0.5138 -0.0888
#>  -0.2445  0.8524 -0.5361  0.6683  0.7929  0.3297  0.0797  0.7173 -0.0883
#>   0.7700 -0.4459  0.3465  0.4964  0.0558 -0.2916  0.3692  0.6004  0.3939
#> 
#> Columns 10 to 18  0.3795 -0.0584 -0.2834 -0.3627 -0.8617 -0.0119  0.4225 -0.3038 -0.2827
#>   0.3420 -0.9069 -0.9504 -0.5479 -0.6312 -0.8850  0.2316 -0.7522  0.3269
#>  -0.2153 -0.0352 -0.7761  0.3365  0.5262  0.2915  0.6184 -0.3115 -0.7324
#> 
#> Columns 19 to 20  0.8671 -0.4008
#>   0.6721  0.4201
#>   0.4237  0.6969
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2108 -0.0776  0.1663 -0.7449  0.0874 -0.2251  0.4585  0.1487  0.1598
#>   0.3216  0.5654 -0.4182 -0.2166  0.0368 -0.1939  0.8003 -0.2707 -0.4724
#>  -0.2808  0.2865  0.4273  0.4748  0.3418  0.1335  0.0332 -0.0746 -0.0099
#> 
#> Columns 10 to 18 -0.4427  0.5719 -0.2913 -0.6432  0.0608 -0.0353  0.6670  0.0409 -0.4220
#>  -0.3498 -0.1199 -0.4163  0.0173 -0.3753 -0.3865  0.4525 -0.5083  0.2381
#>  -0.6749 -0.3813 -0.0491  0.1552  0.0522 -0.0211  0.3255  0.3244  0.4510
#> 
#> Columns 19 to 20  0.4505  0.4030
#>   0.7152  0.1528
#>   0.6807 -0.5319
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3892  0.2749  0.5275  0.5097  0.3637  0.1935 -0.1994  0.3653  0.0847
#>   0.1940 -0.0090  0.0784 -0.2646  0.0969  0.5658 -0.1195  0.4091 -0.0923
#>   0.1141  0.0489  0.5351  0.6561  0.2173 -0.0283  0.6812 -0.2097 -0.3902
#> 
#> Columns 10 to 18 -0.6003 -0.0882 -0.3510 -0.0382 -0.0676  0.3178  0.5852 -0.5205 -0.1465
#>  -0.4371  0.0051 -0.3475 -0.4409 -0.2414 -0.2232  0.2004 -0.6156 -0.1169
#>   0.2128  0.0949 -0.2821 -0.4691 -0.2088  0.3792 -0.0529  0.1971 -0.0177
#> 
#> Columns 19 to 20  0.6124  0.0046
#>   0.7090  0.1474
#>   0.5518 -0.0457
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.2579  0.0183  0.4165  0.2329  0.3193 -0.0827  0.2764 -0.4538 -0.2073
#>   0.4852  0.0612  0.4627  0.4268  0.2030 -0.3537  0.3469 -0.0185 -0.2884
#>   0.3997  0.3534  0.2636  0.4435 -0.1532 -0.0170  0.6431  0.2173 -0.5239
#> 
#> Columns 10 to 18 -0.4395 -0.2320 -0.0705  0.0065 -0.1096  0.1915  0.0548 -0.1749 -0.0466
#>  -0.3663 -0.2056 -0.4269  0.3316 -0.0978  0.1333  0.1061 -0.4058 -0.0297
#>  -0.1721 -0.1629 -0.0940 -0.4658  0.2178  0.1666  0.1170  0.3400  0.1874
#> 
#> Columns 19 to 20  0.5996 -0.4300
#>   0.5550  0.1281
#>   0.5876  0.4094
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.4390  0.1374  0.5626  0.3266 -0.1886  0.1404  0.5180  0.3706 -0.1977
#>   0.4183  0.1509  0.4532  0.5607 -0.1396 -0.0252  0.7135 -0.2397 -0.2068
#>   0.2877  0.2882  0.0316  0.1286 -0.1236  0.4988  0.5019  0.4414 -0.3998
#> 
#> Columns 10 to 18  0.0219  0.1456 -0.1874 -0.2795 -0.0541  0.2475  0.3493  0.0110 -0.2213
#>  -0.0341 -0.2629  0.0980 -0.0503 -0.0734 -0.0179  0.2653  0.1467 -0.0664
#>  -0.1456 -0.1766 -0.0094 -0.4324 -0.2130 -0.0126 -0.0101 -0.0962 -0.1733
#> 
#> Columns 19 to 20  0.6527  0.3782
#>   0.6324 -0.0555
#>   0.5846 -0.1295
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1810  0.2414 -0.2309 -0.1147 -0.2197 -0.7291  0.2320 -0.4298  0.2317
#>  -0.6503  0.6221 -0.4320 -0.1403 -0.2342 -0.8428  0.7683 -0.2143 -0.4128
#>  -0.3418  0.3596  0.0448 -0.0723 -0.6071  0.2664  0.0542  0.1496 -0.5261
#> 
#> Columns 10 to 18  0.2462  0.4074 -0.1022  0.6901 -0.7336 -0.0199 -0.6180  0.1746 -0.3111
#>   0.6978  0.7566  0.0412  0.3416 -0.0530  0.1106 -0.7032  0.8655 -0.3819
#>  -0.0979 -0.2007  0.1701  0.4837 -0.1388 -0.5303 -0.3546  0.5038 -0.3104
#> 
#> Columns 19 to 20  0.3767 -0.3012
#>  -0.3259 -0.0563
#>  -0.5661 -0.2501
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4390  0.1374  0.5626  0.3266 -0.1886  0.1404  0.5180  0.3706 -0.1977
#>   0.4183  0.1509  0.4532  0.5607 -0.1396 -0.0252  0.7135 -0.2397 -0.2068
#>   0.2877  0.2882  0.0316  0.1286 -0.1236  0.4988  0.5019  0.4414 -0.3998
#> 
#> Columns 10 to 18  0.0219  0.1456 -0.1874 -0.2795 -0.0541  0.2475  0.3493  0.0110 -0.2213
#>  -0.0341 -0.2629  0.0980 -0.0503 -0.0734 -0.0179  0.2653  0.1467 -0.0664
#>  -0.1456 -0.1766 -0.0094 -0.4324 -0.2130 -0.0126 -0.0101 -0.0962 -0.1733
#> 
#> Columns 19 to 20  0.6527  0.3782
#>   0.6324 -0.0555
#>   0.5846 -0.1295
#> [ CPUFloatType{2,3,20} ]
#>