RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1047 -0.4828  0.1753 -0.3503  0.2626 -0.7935  0.1124  0.1660 -0.3711
#>   0.6437 -0.4212 -0.3889  0.1212 -0.1341 -0.3928  0.7181 -0.3193 -0.6494
#>  -0.0974  0.9369  0.6158  0.7881  0.2586 -0.0600  0.5977 -0.6533  0.4043
#> 
#> Columns 10 to 18  0.8443  0.5555 -0.1359  0.7832  0.0326  0.4769  0.0985 -0.2051 -0.8520
#>   0.5849 -0.3032 -0.5285 -0.0390  0.2948 -0.7297 -0.0390 -0.5871 -0.0827
#>   0.2774  0.2839 -0.5660 -0.4692  0.3902 -0.0191  0.1941  0.0740  0.9010
#> 
#> Columns 19 to 20 -0.3027 -0.1703
#>  -0.1890 -0.2907
#>   0.7263  0.2793
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0995  0.0618  0.4284 -0.2200  0.1310  0.0954 -0.0801  0.7176  0.0700
#>   0.2249  0.4350  0.2842  0.4380 -0.1182  0.3090  0.3188  0.7658  0.0590
#>  -0.0012  0.6230  0.3103  0.5124  0.1687 -0.0881  0.0303 -0.0342  0.5377
#> 
#> Columns 10 to 18 -0.3123 -0.0667  0.0716 -0.0311  0.2365 -0.0155  0.5680 -0.1763 -0.3624
#>  -0.7303 -0.6504 -0.0839  0.4541  0.3280 -0.2393  0.2838 -0.1846 -0.2000
#>  -0.5242  0.2230  0.3010 -0.5342  0.1225  0.1503 -0.1795 -0.0344  0.5070
#> 
#> Columns 19 to 20 -0.3328  0.1135
#>   0.2492  0.5683
#>   0.9072  0.4128
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3635  0.2011  0.1866  0.5385  0.1060  0.2211  0.4240  0.5330 -0.0048
#>   0.3151  0.2832  0.1278  0.7424  0.2623  0.3712  0.3722 -0.0499 -0.4212
#>   0.1137  0.2631  0.2214  0.1990  0.4311 -0.0191 -0.0637 -0.1201 -0.2803
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4010 -0.4418  0.1072 -0.2715 -0.5487 -0.3392  0.4449 -0.2820 -0.2126
#>  -0.1040 -0.3305  0.1518 -0.1389 -0.7424  0.0929  0.1567 -0.5260 -0.1412
#>  -0.4838  0.0309 -0.7706  0.0984 -0.2308 -0.0356  0.5480 -0.6440 -0.1129
#> 
#> Columns 10 to 18  0.2887  0.3735  0.0769  0.5622  0.3239  0.4727 -0.4142  0.5798  0.2510
#>   0.7442 -0.0698 -0.2696  0.3172 -0.0636  0.0094 -0.0637  0.3219 -0.3233
#>   0.6294 -0.1098 -0.2266  0.3102 -0.5386 -0.3918 -0.7432  0.5278 -0.5182
#> 
#> Columns 19 to 20 -0.2232 -0.2014
#>  -0.3885  0.2251
#>  -0.7054  0.5762
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2272  0.2110  0.3410  0.3930  0.2260 -0.1123  0.5808 -0.3617 -0.4043
#>   0.3960  0.0366  0.2700  0.2097  0.1823  0.2221  0.5875 -0.2822 -0.5571
#>   0.6625 -0.3456  0.4118 -0.0661 -0.0747  0.2392  0.6797  0.2212 -0.3027
#> 
#> Columns 10 to 18 -0.1566  0.1785 -0.1740 -0.1138 -0.3592  0.1634  0.0088 -0.1521  0.2258
#>  -0.1197 -0.2362 -0.4364 -0.0259 -0.2837 -0.0646 -0.2479 -0.4556  0.0040
#>   0.0494 -0.1047 -0.5068  0.0742  0.0451  0.1448  0.5504 -0.4600 -0.6664
#> 
#> Columns 19 to 20  0.8405 -0.2348
#>   0.5672 -0.1867
#>   0.1963 -0.0679
#> [ CPUFloatType{2,3,20} ]
#>