RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3462  0.6600  0.0490  0.2615  0.7503 -0.3242 -0.7344  0.6621 -0.1927
#>  -0.2007 -0.6407 -0.5451 -0.0290  0.4188 -0.8330 -0.8797  0.3688  0.2165
#>  -0.2327  0.6931 -0.7655 -0.4109  0.7192 -0.8587 -0.6812  0.4205  0.8223
#> 
#> Columns 10 to 18  0.6141 -0.3790 -0.1527 -0.5462 -0.6855  0.1419 -0.2691  0.3268 -0.2933
#>  -0.9351 -0.6710  0.3586  0.8037 -0.4948 -0.8530  0.5342  0.0723  0.6970
#>  -0.3095 -0.8533  0.4289 -0.0797  0.4732 -0.7854  0.3805  0.7575  0.6073
#> 
#> Columns 19 to 20 -0.3617 -0.5913
#>   0.6390  0.0549
#>   0.8354  0.7017
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0555  0.0707  0.0730 -0.1434  0.0681 -0.2633  0.0630 -0.5016 -0.2286
#>   0.1768 -0.0749 -0.4518 -0.7712 -0.4524 -0.7514 -0.7672  0.5557  0.0432
#>   0.6085  0.1451  0.2870 -0.6727 -0.3967 -0.0714 -0.1025 -0.1000 -0.5330
#> 
#> Columns 10 to 18 -0.3946 -0.5468  0.6478 -0.1532 -0.4993 -0.6105  0.4586 -0.3833  0.3536
#>  -0.7009 -0.2038  0.7114 -0.3058 -0.5958 -0.5899  0.4106 -0.3160  0.3877
#>  -0.5398  0.4221  0.7574 -0.3651 -0.5049 -0.8075  0.3180 -0.3858  0.5586
#> 
#> Columns 19 to 20 -0.2049  0.7431
#>  -0.2110 -0.2657
#>  -0.3609  0.1398
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2453 -0.4429 -0.0210 -0.2147 -0.4672 -0.2738 -0.4981 -0.2987 -0.0039
#>   0.0145  0.5930 -0.0896 -0.4447 -0.2899 -0.0582  0.0566  0.5615 -0.3270
#>   0.0941  0.1949  0.6015 -0.8385  0.2915 -0.2766 -0.4864  0.3469  0.1521
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3896  0.5425  0.4622  0.3436 -0.0672  0.3894 -0.1629 -0.1238 -0.0492
#>  -0.4879 -0.5715  0.3450  0.1740  0.2889 -0.1023  0.2035 -0.3293 -0.5723
#>   0.5041  0.2659  0.5364  0.3901  0.6119  0.5689 -0.0822 -0.2378  0.2539
#> 
#> Columns 10 to 18 -0.3931  0.4433  0.2539  0.2953  0.1363  0.5613  0.5376  0.1697  0.2208
#>   0.7415  0.4908  0.3029 -0.3857 -0.2433 -0.0113 -0.4873 -0.0667  0.0626
#>  -0.2522  0.1671 -0.7716 -0.0421 -0.5370 -0.3223  0.1403  0.2506  0.2673
#> 
#> Columns 19 to 20 -0.5002  0.0609
#>  -0.8703  0.4923
#>  -0.3994  0.1034
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2275  0.1388  0.0980  0.0347 -0.4488  0.3410  0.1310 -0.1396 -0.0831
#>   0.4467 -0.3428 -0.3329  0.1064 -0.2265  0.1770 -0.0480 -0.6535  0.2153
#>  -0.1246  0.0998  0.4063  0.1219 -0.2054  0.1678  0.5550 -0.0876  0.0479
#> 
#> Columns 10 to 18 -0.4518 -0.1063 -0.0048 -0.2188  0.1454 -0.3842 -0.0262  0.1033  0.0819
#>  -0.0515  0.2013  0.0394 -0.4276 -0.3242 -0.7145  0.0041  0.3193  0.2402
#>  -0.2785 -0.5091  0.0772 -0.2436  0.1496 -0.5473  0.3835  0.4050 -0.3038
#> 
#> Columns 19 to 20 -0.0345 -0.0038
#>  -0.3374  0.2688
#>  -0.1799  0.2891
#> [ CPUFloatType{2,3,20} ]
#>