RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3554  0.9089 -0.6629 -0.9034 -0.1404 -0.8062 -0.6660  0.5626 -0.4017
#>   0.5825  0.1193  0.6063 -0.8213 -0.1917 -0.1991  0.4364  0.8071 -0.3616
#>  -0.6498 -0.5979  0.3241 -0.2432  0.5392  0.6745  0.5592  0.7081  0.3599
#> 
#> Columns 10 to 18 -0.1082 -0.0832 -0.5009 -0.5693  0.3493 -0.1848  0.3445 -0.6257 -0.4626
#>   0.2890  0.6247 -0.5029 -0.4286  0.6584  0.6914  0.7431  0.2113 -0.8500
#>  -0.1676 -0.2118 -0.1390  0.8953  0.5658  0.2742  0.7043 -0.2397 -0.7278
#> 
#> Columns 19 to 20 -0.0495  0.4506
#>  -0.2813  0.3261
#>  -0.3407  0.2392
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2570 -0.0284 -0.3486 -0.0080  0.0654  0.2347 -0.1751  0.4531 -0.3244
#>   0.7081  0.4724 -0.4042  0.4455 -0.2077  0.1997  0.1813  0.6889  0.1362
#>  -0.0334  0.5082 -0.3257  0.5911  0.2652  0.1927 -0.0984  0.3205 -0.0388
#> 
#> Columns 10 to 18 -0.1928  0.2444 -0.0863  0.0784  0.0168 -0.1766 -0.1588 -0.5674  0.1222
#>  -0.0509  0.0090  0.3345  0.6313  0.1908  0.1224  0.5899 -0.6176 -0.2214
#>  -0.1837 -0.2856 -0.2717  0.1817 -0.1975 -0.3334  0.5447 -0.4629  0.0981
#> 
#> Columns 19 to 20  0.3845 -0.1832
#>   0.4530  0.1474
#>   0.3427 -0.0848
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0896 -0.0403  0.3891  0.0809  0.1253  0.0181  0.2481  0.0778 -0.3189
#>   0.1594  0.5248  0.1906  0.5541  0.3344  0.0956 -0.0759  0.4985 -0.0103
#>   0.1211  0.4622  0.3236  0.1956  0.3527 -0.2338  0.4814  0.1369 -0.3576
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1062  0.1054 -0.2549  0.5197  0.6360 -0.2592  0.0814  0.4313 -0.5647
#>  -0.4703 -0.2002 -0.5259  0.0642 -0.1180 -0.4972  0.1195 -0.5728 -0.0934
#>  -0.1016  0.4042 -0.4639  0.1725  0.1088 -0.4093  0.1922 -0.3015  0.3103
#> 
#> Columns 10 to 18 -0.2412  0.0718  0.5416  0.1946  0.0339 -0.2040 -0.3769  0.2037 -0.5132
#>   0.4304  0.1833 -0.5767  0.5275 -0.3248 -0.3063 -0.5696  0.1833 -0.3153
#>  -0.0931 -0.2056 -0.4287 -0.0550 -0.1867 -0.3550 -0.3791  0.0693 -0.3983
#> 
#> Columns 19 to 20  0.2811 -0.3432
#>   0.4568  0.0430
#>  -0.3269 -0.4609
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4743  0.4700  0.1402  0.2625 -0.1491  0.2889  0.0770  0.4726  0.0434
#>   0.3865  0.1883  0.5732  0.2908  0.0394  0.2012  0.5499  0.0385 -0.0932
#>   0.3916  0.2680  0.4453  0.5949 -0.0753  0.4686  0.4848  0.5759 -0.1080
#> 
#> Columns 10 to 18 -0.6051 -0.4260 -0.0319  0.1777  0.1109 -0.3338  0.1480 -0.4728 -0.1406
#>  -0.2557 -0.1046 -0.3736 -0.2561 -0.1792  0.1017  0.3429 -0.2949 -0.3749
#>  -0.0809 -0.0667 -0.4368 -0.4395  0.2061 -0.0255 -0.1013 -0.0215 -0.0635
#> 
#> Columns 19 to 20  0.6425  0.1073
#>   0.5328  0.3450
#>   0.6095  0.5204
#> [ CPUFloatType{2,3,20} ]
#>