RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2845  0.6180 -0.2569 -0.3714  0.5421  0.4232 -0.8909 -0.5584 -0.2123
#>   0.6899 -0.1598  0.6806 -0.3693 -0.5168 -0.6283 -0.2789 -0.7461 -0.1631
#>   0.7246 -0.5993 -0.2005 -0.5483 -0.7241 -0.2158  0.1507 -0.4915 -0.2335
#> 
#> Columns 10 to 18  0.0022 -0.0144 -0.3322  0.5258 -0.7486  0.3653 -0.0041  0.6826  0.2498
#>  -0.5306 -0.5313  0.3778  0.3913  0.6702  0.9476 -0.4393  0.4911  0.5178
#>  -0.4442 -0.4457 -0.0468  0.0596  0.8518  0.1150  0.4445 -0.2295 -0.6581
#> 
#> Columns 19 to 20  0.3539 -0.6989
#>   0.8893  0.0805
#>   0.3269  0.2953
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1868  0.1725  0.0172 -0.0053 -0.4075  0.3888 -0.2543  0.4420 -0.3718
#>   0.5402  0.1380 -0.0434 -0.2671 -0.1165  0.1354 -0.4076  0.2893  0.1355
#>   0.6623 -0.1252  0.1967 -0.0710  0.1591 -0.0474 -0.4752 -0.3594 -0.2518
#> 
#> Columns 10 to 18  0.1230  0.0230 -0.0534 -0.2408  0.3208 -0.3903  0.2990  0.1902 -0.6525
#>   0.2638 -0.1546 -0.0355  0.0839  0.7016 -0.0111  0.1871  0.2258 -0.5215
#>  -0.0448  0.5185  0.2247  0.2917  0.6479  0.1676 -0.0091  0.1759 -0.0879
#> 
#> Columns 19 to 20  0.4186  0.1102
#>   0.0208  0.3703
#>   0.3596  0.1398
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1060  0.2325  0.2673  0.1084 -0.0698 -0.1873 -0.5390 -0.1381 -0.1906
#>   0.2188  0.3311  0.3112  0.5608 -0.2785 -0.7128 -0.0242 -0.2503 -0.1709
#>   0.5839 -0.4745 -0.0004 -0.3944 -0.5993 -0.2181 -0.2966  0.1179 -0.0468
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8341  0.5049 -0.2035 -0.4724 -0.5403 -0.5971  0.5372  0.9282  0.3393
#>  -0.2377 -0.4238 -0.1920  0.5307 -0.3323 -0.1574  0.6637  0.7915 -0.2551
#>  -0.3296  0.2153  0.2631  0.5538 -0.4830 -0.0446 -0.8812 -0.7556  0.1413
#> 
#> Columns 10 to 18 -0.3588  0.6826  0.4777 -0.6471  0.5943  0.4960  0.4423  0.1270 -0.8223
#>   0.8821  0.3082 -0.3428  0.0011  0.2532  0.7623 -0.5887  0.1856 -0.4637
#>  -0.8499  0.2478  0.2599 -0.1656  0.4659 -0.5424  0.6797 -0.6174 -0.3214
#> 
#> Columns 19 to 20  0.1835  0.4828
#>   0.3578  0.5203
#>   0.6979 -0.2169
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0435 -0.3325 -0.3544  0.1769 -0.5431  0.0631 -0.3578  0.0220 -0.2324
#>   0.3553 -0.1615 -0.0899  0.0034 -0.3425  0.0928 -0.7120  0.4041  0.1273
#>   0.3864 -0.5159  0.0152 -0.3065 -0.4631 -0.4079 -0.1053 -0.1417 -0.0778
#> 
#> Columns 10 to 18 -0.0134  0.3158  0.2847  0.3874  0.7087  0.0447 -0.2530  0.4367 -0.2581
#>  -0.2094  0.0853 -0.1553 -0.1763  0.0260  0.0858  0.3700  0.5461  0.0621
#>  -0.4183  0.7017  0.3424 -0.2099  0.7110  0.2006  0.3340  0.1601 -0.2627
#> 
#> Columns 19 to 20  0.1496  0.4786
#>  -0.0233  0.3012
#>   0.3077  0.3330
#> [ CPUFloatType{2,3,20} ]
#>