RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5770 -0.9375 -0.2437  0.6684 -0.6188  0.2441  0.0154  0.0665  0.8005
#>  -0.2125 -0.0950  0.0568  0.9598 -0.2844 -0.9115 -0.5475  0.6721  0.2781
#>  -0.3403 -0.3897 -0.5269  0.2950 -0.9579 -0.6690 -0.1844  0.5781  0.7433
#> 
#> Columns 10 to 18  0.7493 -0.1950 -0.0499  0.3807 -0.3605  0.3109 -0.6333 -0.3431 -0.2938
#>  -0.0109 -0.7119  0.2570 -0.1720  0.0738  0.2329  0.8767  0.2545  0.7118
#>  -0.1046  0.5127  0.3684 -0.1182 -0.0447  0.4536 -0.2672  0.3827 -0.8017
#> 
#> Columns 19 to 20 -0.0819 -0.5967
#>  -0.6845  0.3592
#>  -0.4707  0.5891
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3282  0.0192  0.4493  0.3083  0.2007 -0.1546 -0.2715  0.0814 -0.6225
#>  -0.1192  0.5840  0.2926  0.2000  0.0959 -0.3752 -0.0622 -0.2519 -0.4162
#>   0.1688  0.4793 -0.1459  0.4121 -0.0880 -0.4268  0.5183 -0.1955 -0.1823
#> 
#> Columns 10 to 18  0.1599 -0.3498  0.1827  0.2794 -0.2495  0.0791  0.0565  0.0820 -0.1368
#>  -0.0687  0.4656 -0.2498 -0.6318  0.3328  0.0048  0.3160 -0.3658 -0.1422
#>  -0.4445 -0.3191 -0.1964  0.0900  0.0171  0.3384  0.1812  0.1864 -0.3598
#> 
#> Columns 19 to 20 -0.3424 -0.1822
#>  -0.1109 -0.0809
#>   0.4089 -0.1437
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1652  0.1666  0.3540  0.0042 -0.4083 -0.2578 -0.2720 -0.1801 -0.5814
#>  -0.3368  0.3650  0.1455  0.2307  0.0012  0.0720 -0.0681  0.1613  0.1009
#>   0.0173 -0.2915  0.0375  0.0876 -0.3811 -0.2364 -0.2652 -0.0508 -0.1498
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2197 -0.1134 -0.4527 -0.5994 -0.1647  0.3146  0.2684  0.4867  0.0821
#>   0.1147 -0.2824 -0.2016 -0.7336 -0.7216  0.2367  0.6317  0.7510  0.7139
#>   0.2331  0.0493 -0.0859  0.3845  0.1410  0.7301  0.4356 -0.4282  0.0189
#> 
#> Columns 10 to 18 -0.3131 -0.7286 -0.3789 -0.5728 -0.5351 -0.4083 -0.3017  0.1454  0.1305
#>  -0.4547 -0.5657 -0.3034  0.4009 -0.0994 -0.4418  0.0109 -0.5421  0.3094
#>   0.6965 -0.2139  0.1644 -0.8379 -0.2014 -0.2572  0.0903  0.2775 -0.3724
#> 
#> Columns 19 to 20 -0.2348  0.0856
#>   0.2128 -0.3866
#>  -0.5415  0.4143
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4409  0.4284  0.0317  0.1369 -0.2324 -0.3142 -0.0486  0.0051 -0.0953
#>   0.1329 -0.0383  0.1189  0.0483 -0.4361 -0.5370 -0.2676 -0.1363  0.1310
#>  -0.1186  0.3814 -0.0599  0.0328  0.3294  0.1747 -0.0470  0.3082 -0.4467
#> 
#> Columns 10 to 18 -0.1024  0.3398 -0.1271 -0.0241 -0.2467 -0.2282  0.2583 -0.0887 -0.1923
#>   0.0405 -0.2713  0.2414 -0.2299 -0.2319 -0.2689  0.7827  0.1114 -0.1373
#>  -0.1465 -0.0836 -0.2326  0.2490 -0.1300  0.1167  0.0532 -0.1871 -0.3922
#> 
#> Columns 19 to 20 -0.3582 -0.1802
#>  -0.4208  0.0799
#>  -0.3556  0.1229
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>