RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4266 -0.7983 -0.8833  0.8288 -0.1431 -0.8346  0.4234  0.8001  0.9791
#>   0.2177 -0.6050 -0.8435  0.8745  0.8742 -0.6064  0.7966  0.5190  0.7981
#>   0.3191 -0.7620 -0.1457  0.4093 -0.1802  0.5162 -0.0062 -0.5356 -0.2816
#> 
#> Columns 10 to 18  0.2491  0.7575 -0.6286 -0.0719 -0.8671 -0.1385 -0.2280 -0.4250 -0.6718
#>  -0.6094  0.0717  0.4728  0.3024  0.3235  0.2780  0.3648 -0.8812  0.3218
#>  -0.0383  0.3359  0.3078  0.0381  0.6122  0.5674  0.1341 -0.6203 -0.1248
#> 
#> Columns 19 to 20 -0.0137 -0.7725
#>  -0.4265 -0.3378
#>   0.2545  0.1388
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2926 -0.6913 -0.6471 -0.4287  0.1170  0.4403  0.3585  0.6939 -0.0667
#>   0.8314 -0.4956 -0.1873 -0.1749  0.2257 -0.0025  0.2604  0.4111  0.5933
#>   0.1258 -0.1624 -0.6668  0.3167 -0.2460 -0.0492  0.1592  0.0513  0.4353
#> 
#> Columns 10 to 18  0.1303 -0.1545 -0.1532 -0.1973 -0.0173  0.3845 -0.1680  0.1512 -0.3500
#>   0.0747 -0.3306 -0.4514  0.5602 -0.3467  0.3489  0.0143  0.1853 -0.3422
#>   0.2359 -0.1726 -0.5759  0.2426  0.0230  0.0203  0.0956 -0.5005  0.1159
#> 
#> Columns 19 to 20  0.3884 -0.0639
#>   0.0197 -0.1917
#>  -0.2739 -0.3113
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2910 -0.0729  0.1485 -0.1936  0.1282  0.3456  0.1548  0.1354 -0.1903
#>   0.3572  0.0753 -0.6867 -0.1703 -0.0233  0.2000  0.2082  0.1911  0.1995
#>   0.1273 -0.3910 -0.2824 -0.3686  0.2477  0.2051  0.1003  0.4953 -0.2411
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8699 -0.6849 -0.4887  0.6999 -0.4256  0.2453  0.1329 -0.4768 -0.2027
#>   0.2406  0.2309 -0.7648  0.3028 -0.5163  0.1394 -0.3067 -0.6046  0.3103
#>  -0.2693  0.4270  0.1774 -0.3985  0.4160 -0.6907 -0.2341  0.3860  0.2262
#> 
#> Columns 10 to 18 -0.1609 -0.0139  0.5080  0.7997  0.2288 -0.5738 -0.1370 -0.8074  0.5523
#>   0.6127  0.2748  0.6286  0.6730 -0.2344 -0.1598 -0.0343  0.1184  0.2051
#>  -0.4423 -0.2264 -0.6943  0.0332 -0.2221  0.1493  0.5825 -0.8626 -0.8478
#> 
#> Columns 19 to 20 -0.9416  0.8124
#>   0.0812  0.3753
#>  -0.1817 -0.3872
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5491  0.4723 -0.1177 -0.6926  0.2880 -0.0401  0.1070  0.2306  0.0654
#>   0.3827 -0.1466 -0.1158 -0.7040  0.1352 -0.2070  0.4319  0.5567 -0.2241
#>   0.1884 -0.2231 -0.6579  0.6826  0.1485  0.2562  0.3786 -0.0646  0.3761
#> 
#> Columns 10 to 18 -0.1287  0.0203 -0.4082 -0.1310 -0.8021  0.3254 -0.1633 -0.0463 -0.5170
#>  -0.1855 -0.0209 -0.5467 -0.5588 -0.5652  0.5128  0.0574  0.0059 -0.2016
#>  -0.4460 -0.2323 -0.2581  0.0096 -0.0476 -0.4962  0.3094 -0.2581 -0.0469
#> 
#> Columns 19 to 20  0.2489  0.1113
#>   0.0881  0.1707
#>   0.1087 -0.4602
#> [ CPUFloatType{2,3,20} ]
#>