RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8625  0.2694  0.1712  0.0112 -0.8223 -0.7380  0.8383 -0.1618 -0.8735
#>   0.4923  0.3477 -0.6898  0.1016  0.4747  0.2910 -0.2973 -0.4803  0.2545
#>  -0.7714 -0.1303 -0.3780  0.6444 -0.6392  0.7723  0.5559 -0.8177 -0.1169
#> 
#> Columns 10 to 18 -0.3460 -0.6554 -0.1981 -0.0378  0.0632 -0.0795  0.2943 -0.6280 -0.9842
#>  -0.3273 -0.5725 -0.8724 -0.4413 -0.3580 -0.5572 -0.0728  0.5269 -0.1064
#>   0.0672 -0.0084  0.2945  0.1573 -0.3610 -0.6309  0.9166  0.5763 -0.7731
#> 
#> Columns 19 to 20 -0.1734 -0.6204
#>   0.2348  0.6087
#>  -0.7472 -0.8554
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2843  0.1848  0.2163 -0.0272 -0.0286 -0.0076 -0.0693  0.1776 -0.5597
#>  -0.5032 -0.4148  0.1508  0.5838 -0.3018 -0.2563 -0.1395 -0.1175 -0.2045
#>   0.1447  0.4937 -0.2465 -0.5186 -0.2201 -0.0299  0.0604  0.1195 -0.5552
#> 
#> Columns 10 to 18 -0.1079  0.2390  0.4907 -0.0156  0.1229  0.1981 -0.2573  0.1646  0.8127
#>   0.1709 -0.2949 -0.3829 -0.2250 -0.1603  0.1079  0.0715  0.4012 -0.3893
#>  -0.2423 -0.0140  0.3181  0.4915 -0.0308  0.4011 -0.0676  0.5539  0.4383
#> 
#> Columns 19 to 20 -0.2349  0.0899
#>   0.1816 -0.1987
#>   0.0401 -0.0074
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2416  0.2308  0.3457  0.5154 -0.1128  0.1945 -0.0351 -0.5327 -0.0012
#>  -0.1758  0.2229 -0.1798  0.4380  0.1543  0.4911 -0.1949 -0.0138 -0.2647
#>   0.1306 -0.1169  0.6294  0.2547 -0.2848  0.3927  0.1587 -0.5081 -0.4571
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1516  0.4958  0.5974 -0.3845  0.0476  0.5430  0.4358  0.0078  0.5670
#>  -0.3117 -0.5184 -0.1793 -0.3883 -0.5518 -0.0683  0.2168 -0.3447  0.0508
#>  -0.7347  0.2322  0.2975  0.3485  0.3558 -0.3013  0.7396  0.4179  0.2060
#> 
#> Columns 10 to 18 -0.2218  0.2569  0.1779  0.3215  0.0970  0.4074  0.2223 -0.4438 -0.4219
#>  -0.2234 -0.0015  0.1127  0.0473  0.3058  0.3929  0.1847  0.1970 -0.0503
#>   0.7651  0.1876 -0.4271  0.1047  0.2507 -0.3508  0.4607 -0.2256  0.6829
#> 
#> Columns 19 to 20  0.1051  0.2020
#>   0.4275  0.5847
#>   0.1669  0.9039
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1039  0.2738  0.1644  0.6488 -0.1644 -0.1416  0.0503 -0.1687  0.0082
#>   0.2510 -0.2671  0.4848  0.0722 -0.2680  0.3058  0.4479 -0.0179 -0.2141
#>   0.0418  0.0646  0.3359  0.5422 -0.0649  0.5723  0.3984 -0.6491  0.0513
#> 
#> Columns 10 to 18  0.2846 -0.0103 -0.2501 -0.0129 -0.3194 -0.0758  0.1404  0.0207  0.3264
#>  -0.3238  0.2340  0.0711  0.3595 -0.0772 -0.0965  0.0724  0.3672  0.2874
#>   0.4103 -0.0126 -0.0407  0.4354 -0.3434 -0.0922  0.2884 -0.0970  0.4191
#> 
#> Columns 19 to 20  0.1016  0.1581
#>   0.1399  0.1342
#>  -0.2990 -0.1629
#> [ CPUFloatType{2,3,20} ]
#>