RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5109  0.0352  0.3489  0.8453  0.5257 -0.4545  0.3831 -0.5149 -0.3999
#>  -0.4199  0.2685 -0.5680  0.8242 -0.2048  0.4071  0.0802 -0.1180 -0.1556
#>   0.3168 -0.7954  0.2824 -0.8069  0.0570  0.5347 -0.4381  0.5330  0.0226
#> 
#> Columns 10 to 18 -0.6476  0.1723 -0.6813 -0.5088  0.3698  0.5108  0.7610  0.0624  0.4672
#>  -0.6248  0.1827 -0.3084 -0.7531  0.3459  0.0540 -0.2688  0.1394  0.1015
#>   0.5587  0.1644 -0.7247  0.3492  0.6725 -0.0824  0.4880 -0.3971 -0.2636
#> 
#> Columns 19 to 20  0.5502 -0.5762
#>   0.8394 -0.5525
#>   0.0292  0.8630
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3521  0.4820 -0.2828 -0.0783  0.2374  0.4943  0.0986  0.0599 -0.3382
#>   0.0754  0.2275 -0.0411 -0.4481 -0.1290  0.6512 -0.2247  0.1005 -0.2608
#>   0.0798 -0.2447 -0.5542 -0.2237 -0.2229 -0.1585  0.0752  0.2037  0.2104
#> 
#> Columns 10 to 18  0.4966 -0.1979 -0.4735  0.0153  0.0426 -0.2962 -0.2416  0.3130  0.4183
#>   0.2649 -0.4313 -0.4170  0.0763 -0.0386 -0.1718  0.0250 -0.5625  0.4260
#>  -0.2116  0.1646 -0.2349 -0.5919  0.1634 -0.1744  0.2494  0.3187  0.2133
#> 
#> Columns 19 to 20 -0.0086 -0.2265
#>   0.4279 -0.4875
#>   0.1072  0.2951
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1393  0.3054 -0.4536  0.0727  0.4320 -0.1782  0.3839 -0.3703 -0.3735
#>   0.2932 -0.0426 -0.6986 -0.5783  0.0383  0.5452 -0.1842 -0.0660 -0.5132
#>   0.2524 -0.0607  0.2232 -0.1654  0.2829  0.3058 -0.2209 -0.0899  0.1286
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4712  0.0214 -0.2435  0.3341  0.0567 -0.2674 -0.4516  0.1521  0.5921
#>  -0.1323 -0.4062 -0.0272  0.7361  0.1214 -0.5157  0.2879 -0.3665  0.9141
#>   0.7189  0.6141  0.3327 -0.1738 -0.7700 -0.4990  0.5877  0.6527 -0.5150
#> 
#> Columns 10 to 18  0.3396  0.3192  0.5589  0.5640  0.2260  0.4005 -0.3322 -0.5089  0.7070
#>   0.2468 -0.5298  0.3861 -0.1378 -0.4133  0.8815 -0.5492 -0.8676  0.6270
#>  -0.0206 -0.6681 -0.4405  0.4035  0.4154 -0.3232  0.3305  0.4139  0.6159
#> 
#> Columns 19 to 20  0.0775 -0.7009
#>   0.0218 -0.9282
#>   0.0288  0.4924
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1549 -0.2561  0.1808 -0.6460  0.1561  0.5853 -0.1929 -0.0687 -0.3536
#>   0.1589 -0.7752  0.5302 -0.4977 -0.2845  0.4187 -0.2037 -0.1801 -0.1409
#>   0.2029  0.2740 -0.4327 -0.1174  0.1601 -0.0315 -0.1335 -0.0260  0.2497
#> 
#> Columns 10 to 18 -0.0893 -0.1756 -0.1474  0.3273  0.1793 -0.4647  0.2259 -0.5163  0.2060
#>  -0.4509  0.3680 -0.1131 -0.0341  0.2526 -0.2482  0.6312 -0.1232 -0.1179
#>  -0.3040 -0.0504 -0.1398 -0.7252  0.1079 -0.1781  0.0503 -0.0121 -0.3420
#> 
#> Columns 19 to 20 -0.0596  0.3465
#>  -0.2451  0.6490
#>   0.1278 -0.0964
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>