RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8602  0.7216  0.3266  0.8357 -0.8318 -0.7262 -0.5799  0.3267 -0.8229
#>   0.2916  0.7007 -0.1747  0.2275  0.5811  0.3989  0.8118 -0.4147  0.9549
#>  -0.2039 -0.5070 -0.5266  0.2145  0.0190  0.2925  0.4282  0.0952  0.9035
#> 
#> Columns 10 to 18  0.1591 -0.0737 -0.5804 -0.7511  0.4861 -0.3389 -0.7901 -0.0480  0.8364
#>   0.5735  0.2642 -0.3489 -0.3905 -0.4259 -0.7028 -0.6733  0.3174  0.4279
#>  -0.3629  0.5666 -0.2726 -0.0245 -0.0063 -0.5204 -0.4250  0.4900  0.8270
#> 
#> Columns 19 to 20 -0.5864 -0.4069
#>   0.2587  0.4719
#>   0.4167  0.0462
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0137 -0.1287  0.5749  0.5469 -0.9044 -0.0804 -0.6204 -0.0413  0.0935
#>  -0.6848  0.4890 -0.0942 -0.1103 -0.0217 -0.4984  0.1846  0.3287  0.3348
#>   0.1436  0.0356  0.1660 -0.3226 -0.1395  0.0974 -0.0301  0.3542 -0.1401
#> 
#> Columns 10 to 18  0.9305  0.4168 -0.5855 -0.4370  0.3152 -0.4037 -0.5856  0.5313  0.5544
#>   0.1797 -0.4514 -0.3527 -0.0240 -0.1382 -0.2236 -0.2003 -0.3207  0.4496
#>   0.6126 -0.2820  0.1942 -0.1835 -0.0809 -0.2318 -0.5392  0.4116  0.4101
#> 
#> Columns 19 to 20  0.4298 -0.1336
#>  -0.3980  0.1120
#>   0.1441  0.5668
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3053  0.2568  0.1533 -0.0598 -0.6290 -0.3303 -0.6717  0.0997 -0.4493
#>  -0.2654  0.2218  0.1562  0.6225 -0.2445 -0.3430 -0.1357  0.3488 -0.1540
#>  -0.3478 -0.2184  0.3229  0.5487 -0.1260  0.0456 -0.2855  0.0861 -0.3995
#> 
#> Columns 10 to 18  0.8370  0.1429 -0.3049 -0.2669  0.1932  0.4366 -0.6523  0.4230  0.4680
#>   0.3967  0.0752 -0.2987 -0.4755  0.0148 -0.2945 -0.0849 -0.5490  0.5403
#>   0.3668 -0.1499 -0.2851 -0.2811  0.2466  0.0035 -0.3844  0.1584  0.4427
#> 
#> Columns 19 to 20  0.4681  0.1083
#>  -0.3819  0.2308
#>   0.0151  0.1606
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.3388 -0.2725  0.1452 -0.0834 -0.3342 -0.2455 -0.5233 -0.0978 -0.1853
#>   0.0111  0.2142  0.3291  0.4073 -0.7684  0.0134 -0.2729  0.0240 -0.1758
#>   0.3657  0.3161  0.5496 -0.0741 -0.2450 -0.0835 -0.5178 -0.0820 -0.4004
#> 
#> Columns 10 to 18  0.6758  0.4195 -0.4921 -0.2915  0.1368  0.4970 -0.2517  0.6257  0.3349
#>   0.5671 -0.1759 -0.0776 -0.3747  0.2027 -0.2634 -0.4303  0.1031  0.6133
#>   0.8021 -0.0940  0.2150 -0.3331 -0.0110  0.4765 -0.2428  0.2160  0.3590
#> 
#> Columns 19 to 20 -0.0779  0.0241
#>  -0.1087 -0.2026
#>   0.0155  0.2027
#> 
#> (5,.,.) = 
#>  Columns 1 to 9 -0.3744  0.0714  0.0463  0.3634 -0.6211 -0.2500 -0.2722  0.0371 -0.3010
#>  -0.5140  0.1369  0.3463  0.3656 -0.5993 -0.0989 -0.3265 -0.0147 -0.0154
#>  -0.2268  0.0464 -0.0493  0.3440 -0.5042 -0.1807 -0.4559  0.0087 -0.3569
#> 
#> Columns 10 to 18  0.3764 -0.4206 -0.3947 -0.2861 -0.1350  0.2085  0.1166  0.5903  0.5067
#>   0.5849 -0.1580 -0.6916 -0.4892 -0.0685 -0.1258 -0.2522  0.4786  0.2862
#>   0.3959 -0.0003 -0.5916 -0.3833  0.0314  0.1823  0.0415  0.3136 -0.0097
#> 
#> Columns 19 to 20 -0.2183  0.0309
#>   0.2303  0.1540
#>  -0.3777 -0.1460
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.4455  0.0625  0.1735 -0.2736 -0.2345  0.2214 -0.6394  0.4117  0.2452
#>  -0.3142 -0.4928 -0.3782 -0.6126  0.6180  0.2001  0.4070  0.4264  0.4202
#>   0.1653 -0.3302  0.4631  0.4052 -0.5041  0.6584 -0.0185  0.1668  0.0802
#> 
#> Columns 10 to 18 -0.7638  0.1827 -0.6471  0.3327  0.7053 -0.0893  0.7612  0.2357 -0.5982
#>  -0.5755  0.5408 -0.4381  0.1185  0.6081 -0.4424  0.7388 -0.2726 -0.4224
#>  -0.2852  0.2175 -0.3230  0.6165  0.4056 -0.0945 -0.1662 -0.0628 -0.4486
#> 
#> Columns 19 to 20 -0.0306  0.2791
#>   0.6370 -0.3998
#>   0.1249  0.2034
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3744  0.0714  0.0463  0.3634 -0.6211 -0.2500 -0.2722  0.0371 -0.3010
#>  -0.5140  0.1369  0.3463  0.3656 -0.5993 -0.0989 -0.3265 -0.0147 -0.0154
#>  -0.2268  0.0464 -0.0493  0.3440 -0.5042 -0.1807 -0.4559  0.0087 -0.3569
#> 
#> Columns 10 to 18  0.3764 -0.4206 -0.3947 -0.2861 -0.1350  0.2085  0.1166  0.5903  0.5067
#>   0.5849 -0.1580 -0.6916 -0.4892 -0.0685 -0.1258 -0.2522  0.4786  0.2862
#>   0.3959 -0.0003 -0.5916 -0.3833  0.0314  0.1823  0.0415  0.3136 -0.0097
#> 
#> Columns 19 to 20 -0.2183  0.0309
#>   0.2303  0.1540
#>  -0.3777 -0.1460
#> [ CPUFloatType{2,3,20} ]
#>