RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1654  0.3169 -0.6299  0.1220 -0.5080 -0.7694 -0.0589 -0.6113 -0.7326
#>   0.5777  0.4377 -0.1830  0.5006  0.6698  0.7612 -0.7031 -0.1238 -0.2973
#>  -0.8911 -0.7220  0.3197  0.2920  0.0713 -0.8660  0.5558 -0.4716  0.7452
#> 
#> Columns 10 to 18 -0.6455  0.5393  0.8143 -0.4586 -0.3161  0.3531  0.1046 -0.3861  0.6250
#>   0.6264 -0.8915 -0.4924  0.4300  0.2416  0.6176 -0.3420 -0.4899  0.1617
#>   0.3369 -0.6467 -0.4804  0.0682 -0.2916 -0.4682 -0.5372  0.3669  0.1021
#> 
#> Columns 19 to 20  0.7159  0.1466
#>   0.0855  0.2046
#>  -0.8592 -0.6325
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1680 -0.3160 -0.2572 -0.0985  0.0523 -0.5837  0.2040  0.3568  0.1120
#>   0.0250 -0.7102  0.4946 -0.5195 -0.5597  0.4882  0.2788 -0.5768  0.0568
#>   0.1704  0.1084 -0.2822 -0.4711  0.5507 -0.2080  0.4651 -0.0901  0.0524
#> 
#> Columns 10 to 18  0.2063  0.0366 -0.2522  0.0363 -0.1791  0.2959 -0.4647 -0.2258  0.0273
#>  -0.2787  0.7297 -0.0498 -0.1181  0.4862 -0.5242  0.5449  0.3099  0.0961
#>   0.0991 -0.5058 -0.0113 -0.1015 -0.4756  0.0563 -0.5031 -0.2378  0.3840
#> 
#> Columns 19 to 20  0.0464 -0.1151
#>  -0.3333 -0.6466
#>  -0.2881 -0.1965
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1182 -0.3599  0.0321 -0.0010 -0.0194 -0.2724  0.1735 -0.0323  0.0206
#>   0.3517 -0.0433  0.2100 -0.5191  0.1683 -0.2929  0.5413 -0.4523 -0.3207
#>   0.3131  0.2489  0.1029 -0.2606  0.4298 -0.2642  0.2695 -0.6032 -0.4535
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5750 -0.7317 -0.0718  0.6438 -0.4882 -0.6339  0.1601 -0.4166  0.4242
#>  -0.4393 -0.7810  0.2434 -0.0166 -0.4349 -0.0485  0.1459 -0.1396  0.1541
#>   0.1892  0.6137  0.0943 -0.5202  0.1558 -0.0929 -0.3166  0.1235  0.2424
#> 
#> Columns 10 to 18  0.6729 -0.3163  0.1138 -0.6336  0.3718 -0.0042 -0.3320  0.3123 -0.5416
#>  -0.3226 -0.8276  0.4308 -0.5404  0.4198  0.0083 -0.2996 -0.2202  0.0532
#>  -0.2240 -0.3830 -0.6702  0.5751 -0.7462  0.2531 -0.2360 -0.6843 -0.6326
#> 
#> Columns 19 to 20 -0.1746 -0.5663
#>  -0.4109 -0.2721
#>   0.0895  0.6115
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0101  0.3170  0.0609  0.1082  0.3502 -0.4405  0.4134 -0.0213 -0.6540
#>  -0.1148 -0.1365 -0.2331  0.0528  0.3229  0.1131  0.0511 -0.0638 -0.4471
#>   0.4809 -0.0517  0.0092 -0.6847 -0.3036 -0.1338  0.3197 -0.2649 -0.3757
#> 
#> Columns 10 to 18  0.3640 -0.2291 -0.1406 -0.2721 -0.5904  0.1589  0.0166  0.3962  0.3797
#>  -0.0034 -0.0124 -0.2533 -0.1392 -0.1546  0.2681 -0.4225 -0.3537  0.2558
#>   0.3204  0.6255  0.0722 -0.1437  0.0847 -0.6242  0.0314  0.4609  0.0901
#> 
#> Columns 19 to 20 -0.3361 -0.1349
#>   0.3183 -0.0606
#>  -0.1496 -0.3485
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>