RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7566  0.3281 -0.4263  0.1046  0.8762  0.6939  0.0171 -0.1272 -0.6371
#>   0.4519 -0.2273 -0.5352  0.5679  0.1171  0.1990 -0.6169 -0.4439 -0.0363
#>   0.4993  0.1582  0.3580 -0.4588  0.2067  0.5985 -0.7113  0.5715  0.5013
#> 
#> Columns 10 to 18 -0.8350  0.8258 -0.6430 -0.1286 -0.3757  0.6882  0.4288 -0.4930 -0.5657
#>   0.2836 -0.4910  0.1930 -0.4115 -0.4752  0.4359 -0.0730 -0.4729  0.2240
#>  -0.3374  0.2810  0.6812  0.8212 -0.6924  0.6703 -0.3990 -0.1778 -0.2851
#> 
#> Columns 19 to 20 -0.6130  0.5363
#>  -0.0937  0.3039
#>   0.7471 -0.3927
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5758  0.6122 -0.2221 -0.1802 -0.3546 -0.5037 -0.3241 -0.3730  0.0892
#>   0.2560 -0.5003  0.4681 -0.4879  0.0870  0.5299  0.5459 -0.1652 -0.0725
#>  -0.4797 -0.5482  0.0073 -0.3768  0.3475  0.3377 -0.0836 -0.0299  0.0317
#> 
#> Columns 10 to 18 -0.1721  0.6366 -0.1970  0.4449 -0.4873  0.1417  0.0365  0.1009  0.2675
#>  -0.7429  0.4110 -0.6874  0.4728 -0.2150  0.3101  0.0279 -0.3270 -0.2049
#>  -0.6727 -0.4330  0.3381  0.5100 -0.2782 -0.2527  0.5989  0.7934 -0.3572
#> 
#> Columns 19 to 20 -0.4844  0.1110
#>  -0.3011  0.4557
#>  -0.3121  0.6395
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4166 -0.3090  0.0805 -0.3940  0.2264 -0.0648 -0.3582 -0.2139 -0.2348
#>  -0.1551 -0.0053 -0.0572 -0.2317  0.4156 -0.1628  0.0048 -0.2018 -0.4003
#>   0.1837  0.1502 -0.0883 -0.4840  0.5699 -0.4014 -0.4955  0.4226 -0.2419
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6177 -0.3286  0.2547 -0.9169  0.6200 -0.7964  0.0725 -0.0298 -0.5764
#>  -0.5625  0.4044 -0.7899 -0.7359  0.5422 -0.7242 -0.1042  0.7809 -0.3419
#>   0.5692  0.3596  0.5627  0.8913 -0.3988  0.4031 -0.5342  0.1755  0.3289
#> 
#> Columns 10 to 18  0.2345 -0.0649 -0.2421 -0.4518  0.2458 -0.3828  0.4434  0.4940 -0.2424
#>  -0.6703 -0.7676  0.5549 -0.0222  0.0600  0.5918 -0.5415 -0.5561  0.3721
#>   0.5542 -0.3192  0.6351  0.0155  0.0194  0.6974  0.3816 -0.6076  0.5915
#> 
#> Columns 19 to 20  0.5303 -0.6943
#>   0.7853 -0.5388
#>   0.5147  0.2237
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2426 -0.1878  0.1876 -0.6398  0.3020 -0.2836 -0.5789 -0.3349 -0.3111
#>  -0.1144 -0.5714  0.3568 -0.4574  0.4521  0.1525 -0.0314  0.1068 -0.3500
#>   0.1064 -0.0690 -0.2562 -0.3440  0.3113 -0.2485 -0.4529  0.5326 -0.2276
#> 
#> Columns 10 to 18 -0.6959  0.5612 -0.2181  0.5345 -0.3831 -0.2217  0.5186  0.5707 -0.3299
#>  -0.6487  0.0008 -0.5595  0.2043 -0.5664 -0.2312  0.6521  0.3871 -0.2679
#>   0.4472  0.0692 -0.2860 -0.0655 -0.6095 -0.2518  0.4191 -0.0026 -0.1669
#> 
#> Columns 19 to 20 -0.2714  0.5810
#>  -0.2726  0.7018
#>   0.1299  0.4811
#> [ CPUFloatType{2,3,20} ]
#>