RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2834 -0.7308  0.0635  0.6221  0.3962 -0.1967  0.0103 -0.2310  0.0192
#>  -0.7724 -0.5749  0.4405  0.1354  0.4775  0.7716 -0.8129 -0.2159  0.7662
#>  -0.3264 -0.8062 -0.6281 -0.0667  0.5009  0.6027 -0.1802  0.7278 -0.6312
#> 
#> Columns 10 to 18 -0.5972 -0.5597 -0.3964  0.8580  0.0913 -0.7991  0.8534 -0.1050 -0.6777
#>   0.0440  0.5893  0.8680  0.0506  0.5317  0.3746 -0.2811  0.1678  0.2282
#>  -0.4997  0.1403 -0.2770  0.7699  0.7944 -0.5726  0.6937  0.4407 -0.7409
#> 
#> Columns 19 to 20  0.3655  0.5885
#>  -0.9259  0.9667
#>   0.4955  0.2703
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2125 -0.2388  0.0426 -0.7508 -0.4098  0.4419  0.3190  0.1721  0.6805
#>   0.2165 -0.5824 -0.0082  0.0681  0.6776  0.4280  0.6880 -0.3907  0.5838
#>  -0.2314 -0.5454  0.1104 -0.4373 -0.0173 -0.1169  0.5525  0.1745  0.3597
#> 
#> Columns 10 to 18 -0.2889 -0.4491 -0.1911  0.0139 -0.1175 -0.0464 -0.3218  0.2401 -0.4473
#>  -0.6819  0.1545 -0.5102 -0.1030 -0.0496 -0.7793 -0.1247  0.1857 -0.2940
#>  -0.5777 -0.9042 -0.4746  0.0700  0.5234  0.2339  0.4296 -0.4926 -0.6229
#> 
#> Columns 19 to 20 -0.0923  0.7527
#>   0.1805  0.1677
#>  -0.1830  0.3744
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3812 -0.5674  0.0455 -0.4874  0.2296 -0.3074 -0.3937 -0.1678 -0.4165
#>  -0.5144 -0.1174  0.3589 -0.1460  0.1398 -0.1497 -0.0905  0.4170  0.3760
#>  -0.5904 -0.4297 -0.0646 -0.7894  0.0222  0.0652 -0.0299 -0.1481  0.4502
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4107  0.6278  0.4159  0.3181  0.2437  0.4512  0.6136  0.4837 -0.3397
#>  -0.1815 -0.3516 -0.3232 -0.3768  0.2397  0.3476  0.3970  0.6874 -0.1845
#>   0.7275  0.6027 -0.5002  0.5258  0.3094  0.5731  0.5709 -0.1576 -0.6599
#> 
#> Columns 10 to 18  0.0581  0.2400  0.4154  0.6760  0.5058 -0.1216 -0.2897 -0.0931  0.0564
#>   0.1791  0.6480  0.5418 -0.0718 -0.2955  0.6901 -0.2423 -0.0527 -0.7088
#>  -0.3545 -0.1989  0.0810  0.6410  0.6035 -0.2718  0.2225  0.7567  0.7448
#> 
#> Columns 19 to 20 -0.7556 -0.0199
#>  -0.6733  0.1102
#>  -0.3777 -0.6544
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6061 -0.3619  0.1957 -0.5328  0.1961  0.1000 -0.1427 -0.0166  0.2933
#>  -0.4805 -0.6094 -0.0826 -0.5089  0.0025  0.1681 -0.4042 -0.0786 -0.3363
#>  -0.1035 -0.1079  0.0814 -0.1994  0.6603 -0.0536 -0.0048 -0.3819  0.4475
#> 
#> Columns 10 to 18 -0.4870 -0.1558 -0.6335  0.0362  0.3950  0.2398 -0.1270  0.1310 -0.4550
#>  -0.3720 -0.3144 -0.4828 -0.1904 -0.0460 -0.1934 -0.0140 -0.2220 -0.7539
#>  -0.7592 -0.6341 -0.1003 -0.0040  0.4360  0.1712 -0.1085 -0.1240 -0.2411
#> 
#> Columns 19 to 20 -0.3356  0.5484
#>   0.1991 -0.0857
#>  -0.2994  0.2708
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>