RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1193 -0.0291 -0.1547  0.5655  0.5928 -0.5839 -0.4872  0.1963 -0.1286
#>   0.8373 -0.8354  0.8339  0.6123 -0.5956  0.2284  0.2865 -0.0100  0.3837
#>   0.1021  0.5469  0.0337  0.5255  0.2959 -0.6399 -0.1711  0.6092  0.0463
#> 
#> Columns 10 to 18  0.0655 -0.5540  0.0113  0.0839 -0.1351  0.1842  0.5775  0.7976 -0.3516
#>  -0.8571 -0.4073  0.1162  0.1829  0.7031 -0.4063  0.2992  0.6016  0.1287
#>  -0.9377 -0.2508  0.3667 -0.6093  0.1362 -0.5299  0.7503  0.2407 -0.3589
#> 
#> Columns 19 to 20  0.5765  0.5498
#>   0.4277 -0.2581
#>  -0.7402  0.4464
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1681 -0.1417 -0.2725  0.1793 -0.6105 -0.0549  0.1622  0.2056  0.4309
#>  -0.0307  0.2586  0.1093  0.3774 -0.0762  0.0341 -0.1536 -0.0836  0.3946
#>   0.4292 -0.3753 -0.1960  0.6098 -0.7667 -0.1328  0.2398  0.0439  0.6967
#> 
#> Columns 10 to 18  0.0770  0.1444 -0.2941  0.2560  0.4454  0.3482 -0.2120 -0.3478 -0.1766
#>  -0.2368  0.3290 -0.6256 -0.1032  0.2805 -0.0897  0.7818  0.3435  0.4001
#>   0.0198  0.0496 -0.5970  0.2271  0.7100  0.0188 -0.0661 -0.1254 -0.5431
#> 
#> Columns 19 to 20 -0.1103  0.1283
#>  -0.2138  0.0731
#>   0.0848 -0.0521
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2707  0.2549 -0.3030  0.1115 -0.3496  0.0933 -0.3244  0.3497  0.2291
#>   0.1246 -0.2312  0.0778  0.6123 -0.6498  0.0892  0.1742 -0.2799  0.4960
#>  -0.3063  0.3500 -0.3223  0.5600 -0.3356 -0.0647 -0.0489 -0.0318  0.3677
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1501  0.3139 -0.4667  0.2004 -0.3715  0.5396  0.3176  0.0392  0.6673
#>   0.2560  0.7348  0.2112  0.3187 -0.2371  0.4822 -0.5218  0.0224  0.3349
#>  -0.6401  0.1516  0.1750 -0.1060  0.4536  0.2046 -0.3191 -0.5946  0.2248
#> 
#> Columns 10 to 18  0.6551 -0.0030 -0.4587 -0.3468 -0.1092 -0.4753  0.2193 -0.7295  0.1273
#>   0.4061  0.0725 -0.0890 -0.7023  0.3480 -0.5959  0.0936 -0.2731 -0.0766
#>   0.3813  0.1826 -0.3817 -0.6935  0.1244  0.0649  0.3144 -0.0451  0.2710
#> 
#> Columns 19 to 20  0.4639 -0.6741
#>   0.2387  0.2584
#>  -0.4163  0.1564
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0885 -0.0445  0.1013  0.6825 -0.6100  0.1313 -0.0392  0.0435  0.4039
#>  -0.2655 -0.0318 -0.4176  0.5308 -0.5200  0.1560  0.2403  0.0689  0.5440
#>  -0.1772 -0.2124 -0.3604  0.4889 -0.5359  0.1029 -0.0281  0.2852  0.5727
#> 
#> Columns 10 to 18 -0.7367 -0.1730  0.1010  0.2208  0.2251 -0.0193  0.6146  0.0508 -0.1853
#>  -0.5998 -0.0989 -0.3085  0.4526  0.0950 -0.1097  0.5660  0.2992 -0.3484
#>  -0.3441  0.0792 -0.5843 -0.0214  0.1402  0.4441  0.4317 -0.1773  0.0020
#> 
#> Columns 19 to 20 -0.1730 -0.0494
#>  -0.0300  0.2578
#>  -0.2858  0.0096
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>