RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1897  0.3428  0.1065 -0.1387  0.8293 -0.2745  0.3124  0.6036 -0.6455
#>  -0.1165  0.2345  0.2711 -0.1623  0.7227  0.8827  0.1754  0.6579 -0.5652
#>   0.3261 -0.7965 -0.7921 -0.7203 -0.1727  0.0070 -0.0965  0.1927  0.8349
#> 
#> Columns 10 to 18 -0.6689 -0.7853 -0.4355 -0.6778  0.2076  0.7453  0.5663  0.7017  0.0232
#>   0.0207  0.4132  0.0219  0.1927 -0.5580  0.2035 -0.4782  0.6524  0.5284
#>  -0.2616  0.4594  0.3387 -0.5988  0.5015  0.2417  0.6140  0.7943 -0.1455
#> 
#> Columns 19 to 20  0.8248  0.4595
#>  -0.3957 -0.5767
#>  -0.2274  0.4435
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5671  0.5781 -0.5018  0.0227  0.3206 -0.3438 -0.5381  0.1303  0.0183
#>  -0.7301 -0.3865 -0.6207  0.6235  0.4656  0.2453 -0.2651  0.0078 -0.0931
#>   0.3548 -0.0287 -0.3961 -0.3403  0.2384 -0.3474 -0.1487  0.5716 -0.2161
#> 
#> Columns 10 to 18  0.0289  0.7371 -0.5524  0.0628 -0.0284 -0.6621 -0.2256 -0.5597 -0.3342
#>  -0.3099 -0.0336  0.1470 -0.2459  0.0108  0.3022  0.0294  0.1034  0.4980
#>   0.3243  0.1574 -0.4592 -0.4255 -0.4122  0.0033 -0.0353  0.3223 -0.4172
#> 
#> Columns 19 to 20  0.1600 -0.4239
#>   0.6345 -0.6774
#>  -0.2852  0.0007
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2496  0.6658 -0.2338  0.0980  0.1987  0.3530 -0.4005  0.1783 -0.0966
#>  -0.3409  0.6076 -0.3408  0.1159  0.4343 -0.3835 -0.2018 -0.3212  0.0534
#>   0.1999  0.4734 -0.5170 -0.2299  0.0866  0.2429 -0.6171  0.3472 -0.0212
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2839  0.3762 -0.6600  0.4428 -0.3909  0.0653 -0.1744  0.3945  0.4927
#>   0.6085 -0.1809 -0.3831  0.4868  0.0015  0.3147  0.5886 -0.2424  0.2791
#>  -0.8094 -0.0062 -0.2244 -0.5247  0.6076  0.3103  0.4353  0.1994 -0.1089
#> 
#> Columns 10 to 18 -0.5132  0.7558  0.1895  0.1510  0.6625 -0.2859  0.0850  0.0167 -0.1829
#>   0.1087  0.5130  0.5955 -0.6122  0.4288  0.2089 -0.3254 -0.1279 -0.5371
#>   0.3714 -0.0500 -0.3371  0.4690 -0.5586 -0.0815  0.1202 -0.8323 -0.5482
#> 
#> Columns 19 to 20 -0.0190 -0.2192
#>  -0.4706  0.3312
#>   0.2748 -0.5372
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3555  0.4798 -0.5481 -0.0586 -0.3213 -0.3259 -0.5069  0.3904  0.0740
#>  -0.1763  0.4048 -0.2863  0.1573 -0.1632 -0.0233 -0.5573  0.4347  0.1109
#>  -0.1234 -0.5666 -0.4474 -0.4929  0.0665 -0.0363 -0.1036 -0.1944 -0.1060
#> 
#> Columns 10 to 18  0.5464  0.1771 -0.4996  0.0043 -0.0363 -0.2856 -0.3969 -0.1205 -0.1727
#>   0.5207  0.5317 -0.6245  0.1696 -0.2314 -0.6510  0.1512  0.0843 -0.3447
#>   0.1353  0.2538 -0.0225 -0.5063  0.2122  0.2265  0.3008  0.4669 -0.2995
#> 
#> Columns 19 to 20 -0.1093  0.0336
#>   0.1969 -0.1547
#>  -0.2145  0.0319
#> [ CPUFloatType{2,3,20} ]
#>