RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2632 -0.0898  0.8322 -0.0629 -0.8749  0.4628 -0.6894  0.5985 -0.3149
#>  -0.2680 -0.5896  0.1210 -0.7646 -0.9155 -0.2437 -0.6060  0.0280  0.0477
#>  -0.7021  0.2807  0.1164  0.0352 -0.7092  0.3592  0.3486  0.7993 -0.0474
#> 
#> Columns 10 to 18  0.5225 -0.6259  0.1926 -0.4345 -0.1985 -0.7979  0.0982  0.4493 -0.0543
#>  -0.8532  0.3477 -0.6780 -0.0836 -0.1984 -0.2650  0.1596  0.2376  0.7960
#>   0.3397  0.0086  0.6270 -0.5219  0.5727 -0.4900  0.6672 -0.4929 -0.6327
#> 
#> Columns 19 to 20 -0.6270 -0.2714
#>  -0.6048  0.0067
#>  -0.9327  0.7848
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4614 -0.1713 -0.3151 -0.0717 -0.4417 -0.0130 -0.2635  0.4196 -0.2791
#>   0.6015 -0.4108 -0.0510 -0.0851 -0.5676  0.2884 -0.6447  0.2190 -0.4890
#>   0.2586  0.3240 -0.2526 -0.0309  0.4043 -0.3733  0.3014  0.5146  0.1359
#> 
#> Columns 10 to 18 -0.0650  0.8399 -0.3609 -0.1626 -0.0715  0.3320  0.3346  0.1054 -0.2106
#>  -0.0301  0.4538  0.2444 -0.4227  0.5704 -0.2713  0.3295 -0.0094 -0.8146
#>   0.2121  0.8666 -0.4412 -0.5836 -0.1707 -0.3833  0.0122  0.2179 -0.3228
#> 
#> Columns 19 to 20 -0.1649 -0.3725
#>   0.5854 -0.2467
#>   0.3785  0.3135
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3123 -0.4394  0.2037 -0.3372 -0.4298 -0.1866  0.4185  0.3101 -0.4910
#>  -0.5505 -0.5791  0.1287 -0.2643 -0.1461 -0.0804  0.4596 -0.0350 -0.0097
#>  -0.2543 -0.5329  0.1729 -0.5862 -0.6925 -0.3866  0.3437  0.0511  0.6198
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0006 -0.4270  0.7175 -0.3595  0.4546  0.1297 -0.0020 -0.1555  0.2920
#>   0.3266  0.6604  0.3025  0.1792  0.4039 -0.1941  0.6622  0.5339 -0.6897
#>   0.5298  0.4147  0.0142  0.4268  0.5200  0.3764 -0.2281  0.0146  0.0399
#> 
#> Columns 10 to 18  0.7067 -0.2788 -0.7068  0.7404  0.6126 -0.6512  0.2494  0.4962 -0.2262
#>  -0.0581  0.1076 -0.1053 -0.0713  0.4067  0.3470  0.1246 -0.2507 -0.4706
#>  -0.8150 -0.1400 -0.6387 -0.2088  0.0524 -0.1270  0.3380  0.7492  0.3861
#> 
#> Columns 19 to 20 -0.0876  0.1062
#>   0.1255 -0.3410
#>   0.4045  0.2707
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3488 -0.1189 -0.0652 -0.3287 -0.4750 -0.3325  0.1513  0.6658  0.4468
#>  -0.0037 -0.0371 -0.2145 -0.2569 -0.3434 -0.2114  0.2497  0.2416  0.1236
#>  -0.3725 -0.1056 -0.1532 -0.1447 -0.1483 -0.3513  0.7432  0.2091  0.2094
#> 
#> Columns 10 to 18 -0.0037  0.4643  0.4162 -0.3844 -0.3107 -0.2613  0.2778  0.0397 -0.4459
#>   0.1894  0.5876  0.1677 -0.5018  0.2118 -0.2553  0.2261 -0.2879 -0.1371
#>   0.2207  0.4984 -0.0942 -0.6062 -0.2591 -0.6508  0.4576 -0.3864 -0.2053
#> 
#> Columns 19 to 20 -0.0375  0.3487
#>  -0.2373  0.0629
#>   0.3200  0.5703
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>