RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6794 -0.2517  0.3834 -0.2404 -0.1823  0.5778  0.2863 -0.4726 -0.7161
#>   0.2829  0.6114  0.0959  0.6036  0.7764  0.6819 -0.3810 -0.6734  0.8494
#>  -0.9361 -0.6194  0.7743 -0.5951  0.5540  0.8063  0.2728 -0.8461 -0.3575
#> 
#> Columns 10 to 18 -0.7056 -0.9078  0.1272 -0.4730  0.9370  0.2666 -0.4351 -0.7126 -0.0747
#>   0.8552 -0.0308 -0.5055 -0.3904 -0.1195 -0.9026 -0.7716 -0.5988  0.0881
#>  -0.6510  0.4357  0.5005  0.4040  0.5642 -0.0499 -0.5700 -0.5155 -0.6832
#> 
#> Columns 19 to 20  0.1482  0.5624
#>  -0.6671 -0.3041
#>  -0.6086 -0.1882
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5088  0.0067  0.1886 -0.5277  0.3772  0.3078  0.3617 -0.5570 -0.4516
#>  -0.5705 -0.7070 -0.1260  0.0524 -0.1747  0.5335  0.0718 -0.8887 -0.5012
#>   0.7161  0.1464  0.5794 -0.1369  0.4339  0.3157 -0.0007 -0.2952  0.3595
#> 
#> Columns 10 to 18 -0.3018 -0.3209 -0.2497  0.1610 -0.1141 -0.3618 -0.5150 -0.3456 -0.3259
#>  -0.4387  0.1307 -0.1831  0.5506  0.7289  0.3367  0.3625 -0.1583  0.3406
#>   0.3930 -0.0318 -0.3220  0.4412  0.0984  0.2888  0.3273  0.0205 -0.0331
#> 
#> Columns 19 to 20 -0.0460  0.1246
#>  -0.2423  0.2746
#>   0.0457  0.3293
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.7215  0.7025  0.3033 -0.0383  0.3687  0.3962 -0.4104 -0.6228  0.2940
#>   0.7856  0.3416  0.3919 -0.2786  0.4222  0.6416 -0.1140 -0.2956  0.5249
#>   0.5258  0.2987  0.5629 -0.4510  0.0605  0.3692  0.1870 -0.7893 -0.1431
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3542  0.5130  0.8476 -0.7960  0.1532  0.2235 -0.3919  0.0162  0.3210
#>   0.4997 -0.1483 -0.2409 -0.2962 -0.4688  0.2409  0.0822 -0.6209 -0.1627
#>  -0.2595 -0.2556  0.4142  0.0600 -0.4526 -0.3487 -0.1185  0.2366  0.2416
#> 
#> Columns 10 to 18  0.0755 -0.1666  0.4048  0.6391 -0.6142 -0.0450  0.6146  0.1204 -0.7509
#>   0.0805 -0.5256  0.7176  0.0921 -0.4911 -0.2592 -0.0924 -0.0469 -0.0973
#>  -0.1353  0.6075  0.3007  0.6164  0.5521 -0.1880 -0.0748 -0.6135 -0.5189
#> 
#> Columns 19 to 20  0.0241  0.1901
#>   0.6435  0.3254
#>  -0.4402  0.0542
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0724  0.0174  0.2368 -0.7389  0.4581  0.2524  0.1484 -0.6841  0.1913
#>   0.7413  0.4590  0.7427 -0.1678  0.4820  0.5096 -0.0283 -0.7448  0.4040
#>   0.2419 -0.2137  0.4565 -0.1156  0.5429  0.0919  0.0929 -0.5614 -0.0551
#> 
#> Columns 10 to 18 -0.0658 -0.0397  0.0193  0.2071  0.3562 -0.5562 -0.1084 -0.1205 -0.0369
#>   0.4982  0.1141 -0.5376  0.4516  0.1062 -0.3083 -0.1261 -0.2574  0.2647
#>  -0.5083  0.4540 -0.0639  0.7091  0.0328  0.0999  0.4463 -0.1707 -0.3148
#> 
#> Columns 19 to 20 -0.3872 -0.1700
#>  -0.1191  0.0837
#>  -0.4469 -0.4545
#> [ CPUFloatType{2,3,20} ]
#>