RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1244  0.4514 -0.9031  0.7251  0.5756 -0.5063  0.3968  0.2419 -0.7638
#>  -0.5015 -0.5521  0.2978  0.3717  0.3790 -0.8099 -0.0071 -0.5484  0.4836
#>  -0.7600  0.1426  0.1708 -0.6108  0.6345  0.0812 -0.1938 -0.5141  0.6988
#> 
#> Columns 10 to 18  0.0497  0.4618 -0.6054  0.7753  0.8344  0.7095 -0.2080 -0.2231  0.0581
#>   0.2364 -0.4531 -0.1279 -0.2387  0.1017 -0.1583 -0.3784  0.5549  0.0557
#>   0.6821 -0.5486 -0.7479  0.8729 -0.3526 -0.8120 -0.7874 -0.3411  0.4455
#> 
#> Columns 19 to 20  0.2697 -0.6517
#>  -0.3446  0.0522
#>  -0.7687 -0.1261
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4719 -0.0992  0.1424  0.5432  0.7495  0.3551 -0.0218 -0.2300 -0.0039
#>  -0.5006  0.0935 -0.3045  0.2192 -0.3247 -0.1073 -0.2119  0.1142 -0.1656
#>  -0.5992 -0.4353 -0.2719 -0.7946 -0.3171 -0.3134  0.3021 -0.5129  0.0511
#> 
#> Columns 10 to 18  0.4152 -0.2355 -0.2355 -0.0407 -0.1718  0.3206 -0.6159  0.3942  0.0532
#>   0.2695  0.0187 -0.4704  0.6043  0.4425 -0.1378 -0.1461  0.6158  0.4852
#>  -0.5749 -0.4860  0.1487 -0.0173 -0.3465  0.1775 -0.1400  0.6831  0.5079
#> 
#> Columns 19 to 20  0.3767  0.2967
#>  -0.3940 -0.3107
#>   0.5351 -0.0004
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2550  0.0336 -0.1320 -0.1018  0.5598 -0.2712 -0.0386 -0.6953 -0.1770
#>  -0.5653 -0.1655 -0.2781 -0.3442 -0.0403  0.0890  0.2244 -0.5547  0.0952
#>   0.0519  0.3632 -0.4949  0.1221  0.1618  0.2069  0.0772 -0.0472 -0.0476
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3102  0.1828  0.2421  0.1109 -0.7916  0.0907  0.5551  0.0688 -0.4915
#>  -0.5975 -0.1175  0.2573  0.0328  0.2798 -0.3088 -0.5988 -0.5629 -0.0330
#>  -0.0673 -0.0924  0.4036  0.3639  0.5943 -0.4142  0.6442  0.4906 -0.3320
#> 
#> Columns 10 to 18  0.4725  0.1541 -0.5452  0.5785 -0.6681 -0.4530  0.3128 -0.4016 -0.2021
#>   0.6166 -0.3952 -0.0787  0.2367  0.4817  0.0929  0.4571 -0.0560 -0.1645
#>   0.3246  0.7733  0.3278  0.4756  0.4808 -0.1406 -0.5631 -0.6397 -0.1010
#> 
#> Columns 19 to 20 -0.6498 -0.5064
#>   0.1244 -0.1807
#>  -0.6978  0.1945
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4108 -0.2310 -0.4566  0.1939  0.5904  0.0732 -0.2125 -0.6560  0.0257
#>   0.2737  0.1356  0.0840  0.4508  0.2690 -0.2220 -0.1244  0.0202 -0.3315
#>  -0.0078 -0.0314 -0.2298 -0.2444 -0.1061 -0.0549  0.1830 -0.4945 -0.2753
#> 
#> Columns 10 to 18 -0.1485  0.0048 -0.4362 -0.0043  0.4026  0.0605 -0.0219  0.4924 -0.1003
#>   0.1817  0.0636  0.0590  0.1631  0.0350 -0.0473 -0.5177  0.6273  0.6222
#>  -0.2803 -0.2888 -0.5135  0.1713 -0.2363  0.3404 -0.4818  0.3475  0.1541
#> 
#> Columns 19 to 20 -0.0015  0.6158
#>   0.1109  0.2695
#>   0.3037 -0.6693
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>