RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2056 -0.4858 -0.0706  0.7181 -0.4188 -0.7746 -0.6381  0.4719  0.4911
#>   0.7070 -0.1493 -0.4521 -0.5958 -0.6574 -0.4898  0.8783 -0.8318 -0.7861
#>   0.8737 -0.3854  0.7877  0.6181  0.1996  0.4020  0.2586  0.4936 -0.7710
#> 
#> Columns 10 to 18  0.3822  0.7473 -0.7279 -0.8531 -0.3816  0.1935 -0.0360 -0.1254  0.5253
#>   0.7342  0.5548 -0.2886  0.4092 -0.0244 -0.9163 -0.4852  0.2199  0.8092
#>  -0.1605 -0.4909  0.2296 -0.0174  0.3013 -0.0895 -0.3354 -0.3589  0.5500
#> 
#> Columns 19 to 20  0.7159 -0.7369
#>  -0.7281 -0.4229
#>  -0.7807  0.7549
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3635  0.0159  0.0673  0.0874  0.1194 -0.7130  0.4227 -0.3343 -0.4224
#>   0.2346 -0.0578  0.5541  0.4427 -0.1319  0.2452  0.3055  0.1563 -0.1295
#>   0.4761  0.4590  0.4514  0.3303  0.0765 -0.3042 -0.2005  0.7752  0.2358
#> 
#> Columns 10 to 18  0.6812 -0.2190  0.2712  0.4194  0.3222 -0.5791  0.2270  0.1311  0.4450
#>   0.3736 -0.5766  0.4931 -0.3591 -0.2934 -0.1460 -0.5327  0.3438  0.0213
#>  -0.0983  0.2144 -0.3376 -0.1649  0.2992  0.1891 -0.4653  0.3297  0.2594
#> 
#> Columns 19 to 20 -0.6813 -0.4692
#>  -0.8194  0.4267
#>  -0.4871  0.2288
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0894  0.1146  0.0100  0.0510  0.0524  0.2959  0.4220  0.1577  0.0689
#>   0.6759  0.1868  0.2251  0.1462 -0.0492  0.1399  0.1692  0.2209 -0.2117
#>   0.7545  0.1797  0.5655  0.2790  0.0255  0.0366  0.0287  0.1001 -0.2537
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1778 -0.3162 -0.7069  0.8644  0.7441 -0.0238  0.6180  0.3746 -0.0384
#>  -0.1602  0.0457 -0.3401 -0.5666  0.2514  0.3993  0.1667  0.2681 -0.3980
#>  -0.2670  0.4106  0.2690 -0.0772  0.1251  0.0846  0.2678  0.3853 -0.3599
#> 
#> Columns 10 to 18  0.7637 -0.5055 -0.2528  0.5607  0.0836 -0.2382  0.2748 -0.4034 -0.6328
#>  -0.1992  0.3511 -0.4473 -0.1369  0.7823  0.2539  0.3449 -0.2375 -0.1489
#>   0.4740  0.6437 -0.6631  0.8381  0.4215  0.5295 -0.4571 -0.3493 -0.1728
#> 
#> Columns 19 to 20 -0.4567 -0.4656
#>   0.4056  0.2861
#>  -0.2635  0.0548
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0374  0.1388  0.0432  0.5308  0.1954  0.1060  0.1497 -0.0036  0.1233
#>   0.6958  0.3962  0.4643  0.2338 -0.4176  0.0333  0.1984  0.3928 -0.0797
#>   0.4256  0.0408  0.3412  0.5911  0.1169  0.0150  0.0670 -0.1796 -0.1240
#> 
#> Columns 10 to 18 -0.3538 -0.2698 -0.6043 -0.4541  0.7057 -0.5980 -0.4127 -0.3345  0.6860
#>   0.0711  0.3851  0.1794 -0.1113  0.1730 -0.0453 -0.4358 -0.3114  0.4109
#>   0.4833  0.0331 -0.1783 -0.0996  0.1363 -0.0322 -0.5356 -0.0688  0.1907
#> 
#> Columns 19 to 20 -0.7332 -0.2233
#>  -0.5055  0.2815
#>  -0.6701 -0.0512
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>