RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5582  0.0397 -0.0449 -0.4160 -0.6328 -0.5094  0.5065  0.1150  0.0079
#>  -0.0656  0.3527 -0.5777  0.0775  0.0840  0.0546  0.7311  0.1984 -0.0167
#>  -0.0739 -0.0865  0.2299  0.2034 -0.8908 -0.6398  0.4517 -0.2925 -0.6285
#> 
#> Columns 10 to 18  0.2616 -0.1598  0.2822  0.0667 -0.2189 -0.1386  0.4091 -0.6930 -0.5737
#>   0.1622 -0.8221  0.0040 -0.6636  0.5545 -0.5635  0.4342  0.5156 -0.1576
#>  -0.8237  0.2640  0.7206  0.6539  0.5347  0.5901  0.1680  0.1478  0.5727
#> 
#> Columns 19 to 20 -0.1412  0.0417
#>  -0.3894  0.8152
#>  -0.1565  0.0430
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2658  0.2771 -0.2892  0.3211 -0.0310  0.6179  0.3513 -0.0748 -0.3298
#>  -0.4902 -0.2239  0.4655 -0.3593  0.0218 -0.3145  0.2839  0.0971  0.5609
#>  -0.3875  0.4418  0.0119  0.3547 -0.3990  0.1096  0.0626 -0.4597  0.1249
#> 
#> Columns 10 to 18 -0.0075  0.5362 -0.0286  0.1357 -0.2759  0.3308  0.3670  0.4546 -0.4260
#>  -0.1334 -0.4990  0.2438  0.4781 -0.3655  0.7383  0.1467 -0.1599 -0.7514
#>   0.8135 -0.0861 -0.3568  0.0621 -0.3707 -0.1137 -0.1903 -0.2292 -0.6020
#> 
#> Columns 19 to 20 -0.1955  0.3271
#>   0.3761  0.6635
#>  -0.0684  0.0790
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.6917  0.2197 -0.5037  0.6215  0.0693  0.0872  0.0655  0.2776  0.3775
#>  -0.2395 -0.1612  0.2283 -0.0229  0.4362 -0.0154  0.5324  0.1658 -0.1916
#>  -0.6565  0.6117 -0.5956  0.4769  0.2051  0.3091 -0.1538  0.3814  0.2422
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7025 -0.1812 -0.3551 -0.5923 -0.4531  0.6206 -0.5246 -0.1782  0.2413
#>   0.4080 -0.0420  0.1738 -0.4410 -0.2361  0.3021  0.6673 -0.4209  0.6092
#>   0.3513  0.1815  0.7000 -0.6954 -0.3633 -0.5652 -0.0206  0.2699 -0.2573
#> 
#> Columns 10 to 18  0.0419 -0.5389  0.4273 -0.1425 -0.3218 -0.4722 -0.0318 -0.2279  0.1593
#>   0.3757  0.3989  0.6713  0.0529  0.3456  0.2754  0.3735 -0.0811 -0.3995
#>  -0.7708 -0.6206 -0.2158  0.0182  0.2238  0.0428 -0.2491  0.0407  0.2119
#> 
#> Columns 19 to 20 -0.5217  0.1413
#>  -0.0979 -0.2322
#>  -0.2941 -0.4250
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0036  0.0533 -0.0674  0.1223  0.2132 -0.0633  0.1705 -0.0588  0.0546
#>  -0.3874  0.0488 -0.0429  0.2348  0.3246  0.0912  0.3051  0.4103  0.0791
#>  -0.2190 -0.1282 -0.0526  0.2288 -0.1732 -0.0491  0.3128 -0.5271 -0.2257
#> 
#> Columns 10 to 18 -0.0180  0.0731  0.1258  0.1935 -0.1153  0.3208 -0.0417 -0.2216 -0.4037
#>   0.0934  0.2395  0.0786  0.3210 -0.5669  0.6218  0.1017  0.0838 -0.4808
#>   0.1429 -0.2541  0.1719 -0.0826  0.2177 -0.1566 -0.1133  0.3168 -0.4220
#> 
#> Columns 19 to 20  0.0862  0.3528
#>   0.3815  0.3242
#>  -0.1171  0.3360
#> [ CPUFloatType{2,3,20} ]
#>