RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1934 -0.8273  0.4201  0.3753 -0.0196 -0.0706 -0.8199  0.5658  0.5612
#>  -0.3378 -0.7253 -0.0129  0.1692  0.7948 -0.8753 -0.9178  0.7665  0.6192
#>   0.1122  0.4957 -0.5846 -0.3318 -0.4358  0.4028 -0.6117  0.4385 -0.4668
#> 
#> Columns 10 to 18  0.6453  0.2012  0.3557 -0.8061 -0.3981  0.9358 -0.5060 -0.6350  0.0407
#>  -0.1196 -0.5152  0.8548 -0.6717 -0.2021  0.4229  0.2767 -0.5877 -0.2890
#>  -0.8190  0.1032 -0.6128  0.6441  0.0750 -0.8525  0.1290  0.6725 -0.1300
#> 
#> Columns 19 to 20  0.8265 -0.7547
#>  -0.1561 -0.7478
#>  -0.5494 -0.4401
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5529  0.1005 -0.0260  0.3822 -0.1527  0.1600  0.1641  0.6074 -0.2487
#>  -0.4158 -0.2223 -0.2718  0.5091  0.4526  0.0008 -0.4911  0.2757  0.1036
#>   0.6144 -0.0237  0.4160 -0.2719 -0.1494  0.3226 -0.4831  0.0710  0.0189
#> 
#> Columns 10 to 18 -0.6935 -0.3683  0.4729 -0.1150  0.3175 -0.7106 -0.2192  0.3493  0.4045
#>  -0.0983 -0.4678  0.2435  0.2904  0.1967 -0.4151 -0.4284 -0.5967 -0.4303
#>  -0.6052 -0.2918  0.5690 -0.3736 -0.6409  0.5029 -0.0559  0.0154  0.1140
#> 
#> Columns 19 to 20 -0.3180 -0.7145
#>   0.3497 -0.5278
#>  -0.4312 -0.1444
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3798  0.5265 -0.0140  0.5068 -0.1901 -0.1188 -0.5014  0.6453 -0.4647
#>  -0.2728 -0.4269 -0.1474 -0.2278 -0.0999  0.2445 -0.6157  0.6076 -0.2459
#>  -0.3576 -0.2757 -0.2510  0.4158  0.0732 -0.1790  0.0256  0.5820  0.2503
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1734  0.0734 -0.7644  0.2241  0.7694  0.2525 -0.4554  0.6840  0.1605
#>  -0.5350  0.0393  0.1419 -0.4632  0.2203 -0.0187 -0.2630  0.2281  0.2643
#>  -0.4813  0.6105  0.6627  0.4030 -0.5001 -0.1843  0.6279 -0.0143  0.0736
#> 
#> Columns 10 to 18  0.0954  0.5588 -0.0825 -0.4533 -0.3229  0.4258  0.3537 -0.1699  0.3413
#>   0.1889 -0.0570 -0.0701 -0.4696 -0.4723 -0.1010  0.1084 -0.2656 -0.5850
#>  -0.6336  0.0669  0.1952 -0.4073 -0.2782 -0.2017  0.2648  0.1471  0.2160
#> 
#> Columns 19 to 20 -0.6661  0.1673
#>  -0.5621  0.3521
#>  -0.3353  0.6931
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2216  0.1659 -0.3707  0.2336  0.1854 -0.1515 -0.4331  0.4971  0.0035
#>  -0.3291 -0.1378 -0.3268  0.3501 -0.1975  0.1682 -0.5118  0.3716 -0.0627
#>  -0.3308 -0.2092 -0.4917  0.2017 -0.4923  0.5881 -0.1150  0.5551 -0.3432
#> 
#> Columns 10 to 18  0.0050 -0.2427 -0.3273 -0.2260 -0.0383 -0.0732  0.1357  0.0611 -0.3958
#>  -0.0680 -0.1296 -0.0651 -0.0604  0.0036 -0.0150 -0.3559 -0.0073 -0.3588
#>  -0.2101  0.4996 -0.2463 -0.0492  0.1510  0.0329 -0.1156  0.3853 -0.3023
#> 
#> Columns 19 to 20  0.3211 -0.4945
#>   0.1821 -0.5293
#>   0.1196 -0.5901
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>