RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8810  0.5478  0.4186 -0.7565 -0.4100 -0.1953 -0.1274 -0.3879  0.1289
#>  -0.5118 -0.3974 -0.0194 -0.5645  0.0401  0.3434 -0.1665 -0.4827 -0.4815
#>  -0.4541  0.3839 -0.1711 -0.2238  0.3539  0.4603 -0.9051 -0.7320 -0.5433
#> 
#> Columns 10 to 18  0.4729 -0.9109  0.0969 -0.6269  0.2792  0.7162 -0.3465 -0.8279 -0.7099
#>   0.2432 -0.7500  0.5093 -0.2257 -0.3410  0.6972 -0.4531 -0.2967 -0.1498
#>   0.4857 -0.4728  0.2252  0.4490 -0.7612  0.3359  0.7097  0.2921  0.3692
#> 
#> Columns 19 to 20  0.5717 -0.0856
#>   0.3733 -0.3975
#>   0.4353 -0.0908
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5372  0.3606 -0.0998 -0.5708  0.1647  0.8945 -0.1060 -0.2229 -0.6489
#>  -0.2963 -0.1445  0.1149 -0.3084 -0.3292  0.3885  0.3598 -0.2306  0.1950
#>   0.2913  0.2698  0.3698 -0.0511 -0.2896  0.1210  0.3611 -0.0756  0.3520
#> 
#> Columns 10 to 18 -0.0939 -0.6265  0.2738  0.7183 -0.0602  0.0590  0.2293 -0.4012 -0.3784
#>   0.2763 -0.7689  0.0364  0.4846 -0.1743  0.3847 -0.5496 -0.3844 -0.5579
#>  -0.0544 -0.6941  0.1009  0.5540 -0.0058  0.3185 -0.4698 -0.7545 -0.0896
#> 
#> Columns 19 to 20 -0.2788 -0.3564
#>  -0.3464 -0.2582
#>   0.2182 -0.2828
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.6642 -0.2573  0.4519 -0.1566  0.0770 -0.3539  0.4038 -0.3049 -0.0796
#>   0.3532  0.2908  0.2537 -0.1380  0.2578  0.2971  0.2873  0.0245 -0.5384
#>  -0.6464 -0.0310  0.1482 -0.1600 -0.1822 -0.1551  0.2818 -0.6332  0.1809
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3475 -0.9092 -0.1829 -0.3871 -0.4423 -0.2552 -0.2000  0.6174 -0.5977
#>  -0.6584 -0.6653  0.7649  0.5629  0.5312  0.0348  0.0867  0.1311  0.3989
#>   0.4510 -0.7302  0.4779  0.3318 -0.6744  0.2471 -0.1119 -0.1118  0.7759
#> 
#> Columns 10 to 18 -0.4760  0.1969  0.5176  0.0764 -0.1500  0.0250  0.5194  0.2195 -0.2015
#>   0.6478 -0.0629  0.6470 -0.2539 -0.4713  0.5675  0.1222  0.2189  0.2739
#>  -0.8724  0.7726  0.6100 -0.3422 -0.6395 -0.7525 -0.2410 -0.1510 -0.1347
#> 
#> Columns 19 to 20  0.7504  0.4710
#>   0.3038  0.2556
#>  -0.4125  0.6189
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1994  0.4039  0.4018 -0.2259  0.1656  0.1027  0.2745 -0.6085 -0.4786
#>   0.5440  0.2542  0.4498  0.5532  0.1763  0.0887  0.5915 -0.1978 -0.0066
#>  -0.6302 -0.0405 -0.2081 -0.6044 -0.0263  0.5500  0.0789 -0.3388 -0.1213
#> 
#> Columns 10 to 18 -0.1730 -0.6424 -0.1053  0.5285 -0.1606  0.3343  0.0974 -0.0746  0.3089
#>  -0.4329 -0.5327  0.4063  0.3839 -0.3806 -0.0868 -0.0911 -0.6082  0.1647
#>  -0.0131 -0.5948 -0.2887  0.4032 -0.0382  0.5998 -0.3701 -0.6872 -0.3268
#> 
#> Columns 19 to 20 -0.3442 -0.2842
#>   0.0349 -0.5793
#>  -0.2187 -0.3856
#> [ CPUFloatType{2,3,20} ]
#>