RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4808 -0.8046 -0.0271  0.6247  0.4566  0.1426  0.4181  0.3714  0.4079
#>  -0.9654  0.8005  0.1523  0.2327  0.1605 -0.0447  0.3670  0.3074 -0.4857
#>   0.1578 -0.5759  0.1188 -0.1582  0.7587 -0.4034 -0.2996 -0.5426 -0.7094
#> 
#> Columns 10 to 18  0.5534  0.2811 -0.2366  0.2582 -0.1586  0.5385 -0.5783  0.0564 -0.2020
#>   0.2520 -0.6209  0.6783 -0.1543 -0.7738 -0.2341  0.7532  0.4779  0.7336
#>   0.9193  0.5560  0.2204 -0.4982 -0.7377 -0.0011 -0.0686  0.8294  0.0149
#> 
#> Columns 19 to 20 -0.2553  0.4185
#>   0.9307 -0.6030
#>   0.0220  0.7032
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1440 -0.2005  0.5385  0.1114 -0.4436  0.1136  0.0597  0.0963 -0.4874
#>   0.1935 -0.5292  0.2816 -0.6835  0.1176 -0.2717 -0.2604  0.4316 -0.1604
#>  -0.4929 -0.4562  0.4894  0.4128 -0.0544 -0.2061  0.5951  0.4191 -0.3842
#> 
#> Columns 10 to 18  0.0693  0.2889 -0.0192  0.0490  0.1041 -0.2422  0.1624  0.0021 -0.2775
#>   0.1209 -0.3311  0.2196 -0.3705 -0.5610  0.4981 -0.0854  0.3929 -0.1083
#>  -0.3222 -0.1619  0.0807  0.1913 -0.2335 -0.0752  0.1061 -0.1805  0.2158
#> 
#> Columns 19 to 20 -0.1660  0.3970
#>  -0.2012 -0.1669
#>   0.6463  0.2377
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0320 -0.3983  0.3921  0.0248  0.0965  0.2755 -0.4601  0.1141 -0.2487
#>   0.3599 -0.0820  0.2549 -0.1456  0.0388 -0.0404 -0.1458  0.4337 -0.4659
#>   0.5489 -0.2757 -0.0789  0.2679 -0.1026 -0.5454 -0.3125  0.2471 -0.1475
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4237 -0.8358 -0.9592  0.5557  0.1625  0.8180 -0.1515 -0.7931  0.0031
#>   0.0615  0.0325  0.7321 -0.3204  0.1812  0.0454 -0.1680 -0.3618  0.3300
#>  -0.1035 -0.0138  0.2094  0.1600 -0.5907 -0.5261  0.0256  0.3092  0.1509
#> 
#> Columns 10 to 18 -0.0661 -0.5657 -0.2145 -0.9146 -0.1497 -0.0595 -0.0445  0.8094  0.3448
#>   0.8239 -0.3305 -0.1172  0.1565  0.2573 -0.1147 -0.1861  0.1308 -0.3385
#>   0.3857 -0.7423  0.3844  0.0744  0.1626  0.3168  0.6250  0.4253 -0.2980
#> 
#> Columns 19 to 20 -0.3839  0.4125
#>  -0.6497 -0.5311
#>  -0.7454  0.2527
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1718 -0.4557  0.0887  0.2103  0.3563 -0.4043 -0.3892  0.5659  0.0184
#>  -0.3395  0.1039  0.4695  0.2881  0.0178 -0.0170  0.2387 -0.1720 -0.2063
#>   0.4692 -0.0978  0.1212  0.2254 -0.4219 -0.0843 -0.3830  0.2167 -0.2637
#> 
#> Columns 10 to 18  0.7519  0.3948  0.0525 -0.0346 -0.4577  0.4459 -0.3093  0.7701 -0.3036
#>   0.1648 -0.4498 -0.2595  0.1686 -0.3541 -0.1308 -0.0066  0.2184 -0.2900
#>   0.4507 -0.2798  0.1199  0.1075  0.0055  0.1892 -0.4412  0.1282 -0.4879
#> 
#> Columns 19 to 20  0.0516  0.3269
#>  -0.2031  0.0609
#>  -0.1214  0.1211
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>