RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6186  0.5620  0.7404  0.8593 -0.7473 -0.7471 -0.3121 -0.2555 -0.1065
#>  -0.7451 -0.7108 -0.8344 -0.4518  0.3603 -0.0396  0.9143 -0.5399  0.6111
#>  -0.2576  0.5366 -0.1994  0.6793  0.4971  0.0459  0.1755 -0.1373  0.1454
#> 
#> Columns 10 to 18  0.4095 -0.7838 -0.3927  0.2682  0.1770 -0.7025 -0.0463  0.5617  0.5747
#>   0.3108  0.6509 -0.1387 -0.0166  0.5915  0.3005 -0.7740  0.7948  0.7589
#>   0.0880  0.6957  0.2646 -0.3136  0.6575  0.2860  0.6904  0.5805  0.5319
#> 
#> Columns 19 to 20 -0.5697 -0.4914
#>   0.7106  0.9561
#>  -0.3786  0.3646
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4910  0.4318  0.4423  0.3573 -0.6975 -0.0807 -0.6504  0.1805  0.5306
#>  -0.5600  0.3068 -0.1516 -0.0911 -0.2027  0.5094 -0.4296 -0.0564 -0.1409
#>  -0.5505  0.4791 -0.1759 -0.0279 -0.3931  0.2407  0.0634 -0.0066  0.3796
#> 
#> Columns 10 to 18  0.8368  0.4791 -0.2245  0.1005  0.0441 -0.3326 -0.3432  0.4337  0.0174
#>   0.5020 -0.5018 -0.0141 -0.2222  0.2136 -0.1004 -0.0757  0.7020  0.0351
#>   0.6500 -0.2661 -0.2985 -0.4226  0.4401 -0.4765 -0.0787  0.3960 -0.3089
#> 
#> Columns 19 to 20  0.6985 -0.2562
#>  -0.2929  0.5234
#>  -0.5384  0.3232
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2650  0.0645 -0.1409  0.4577 -0.4091 -0.3017 -0.2426  0.1873 -0.6370
#>  -0.2652  0.3249  0.1470  0.2524 -0.0570  0.1481 -0.3386  0.0505 -0.1306
#>  -0.3391  0.3288  0.0400  0.4009 -0.2990 -0.3598  0.0613  0.1979  0.2308
#> 
#> Columns 10 to 18  0.4149  0.0259 -0.3793 -0.1469  0.5203  0.4130 -0.4176  0.1940  0.5770
#>   0.2503  0.3012 -0.4300 -0.5031  0.3391  0.0196  0.4427  0.3464  0.1534
#>   0.5764  0.3218 -0.4014 -0.5258  0.3474  0.0379 -0.0492  0.0760  0.5603
#> 
#> Columns 19 to 20  0.4371  0.2214
#>   0.1106  0.1027
#>   0.2012  0.0961
#> 
#> (4,.,.) = 
#>  Columns 1 to 9 -0.2165  0.0918  0.6146  0.0715 -0.2452 -0.0942 -0.4550 -0.2641 -0.1897
#>   0.0472  0.1431 -0.1655 -0.0666 -0.5826 -0.0451 -0.3385 -0.0334  0.3197
#>  -0.4806  0.2759  0.0300  0.2951 -0.5265  0.0246 -0.1889 -0.0296 -0.1879
#> 
#> Columns 10 to 18  0.6209 -0.2944 -0.0214  0.1030  0.0241  0.4961 -0.1119  0.3791  0.6579
#>   0.6536  0.1136 -0.2312 -0.2150  0.2321 -0.1876 -0.3432  0.2453  0.3254
#>   0.5802 -0.0957 -0.3856 -0.3560  0.3004 -0.1052 -0.2435  0.1394  0.2513
#> 
#> Columns 19 to 20 -0.0260  0.1232
#>  -0.3973 -0.1018
#>  -0.2458  0.2102
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.4000  0.0685  0.3202 -0.1512 -0.1905  0.0851 -0.6651  0.1486 -0.2440
#>   0.3247  0.0325  0.1817  0.2824 -0.4737 -0.5553 -0.1187  0.2275 -0.1896
#>  -0.2652  0.2843  0.0448  0.4260 -0.6998  0.1537 -0.5185 -0.1683 -0.2697
#> 
#> Columns 10 to 18  0.5710 -0.2001 -0.3588 -0.3787 -0.0693  0.2681  0.2900  0.2213  0.3867
#>   0.7652  0.3905 -0.3227 -0.3664  0.1635  0.3607 -0.5696  0.4727  0.5143
#>   0.6653  0.0671 -0.3229 -0.4159  0.4661 -0.1779 -0.2982  0.3463  0.4009
#> 
#> Columns 19 to 20 -0.2047  0.0243
#>   0.4411  0.2155
#>   0.0950  0.0726
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.2492  0.0591  0.1429  0.0282 -0.0398  0.6613 -0.5157 -0.1438 -0.3660
#>   0.4573 -0.8330 -0.1958  0.1748  0.1753  0.1947 -0.1106  0.0371 -0.1961
#>   0.5888  0.2264 -0.3649  0.0609 -0.6084 -0.2117 -0.3560 -0.1886  0.1761
#> 
#> Columns 10 to 18 -0.0827 -0.2866  0.7160 -0.0438  0.4008  0.3939  0.4146 -0.8194  0.2837
#>  -0.0865 -0.3350 -0.1190 -0.4787 -0.2445  0.4084 -0.3121 -0.2374  0.2993
#>  -0.0722 -0.4958  0.2634  0.3816  0.4413 -0.1697 -0.0929 -0.0662 -0.2577
#> 
#> Columns 19 to 20 -0.2204  0.6337
#>   0.5901  0.0879
#>   0.8401 -0.1298
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4000  0.0685  0.3202 -0.1512 -0.1905  0.0851 -0.6651  0.1486 -0.2440
#>   0.3247  0.0325  0.1817  0.2824 -0.4737 -0.5553 -0.1187  0.2275 -0.1896
#>  -0.2652  0.2843  0.0448  0.4260 -0.6998  0.1537 -0.5185 -0.1683 -0.2697
#> 
#> Columns 10 to 18  0.5710 -0.2001 -0.3588 -0.3787 -0.0693  0.2681  0.2900  0.2213  0.3867
#>   0.7652  0.3905 -0.3227 -0.3664  0.1635  0.3607 -0.5696  0.4727  0.5143
#>   0.6653  0.0671 -0.3229 -0.4159  0.4661 -0.1779 -0.2982  0.3463  0.4009
#> 
#> Columns 19 to 20 -0.2047  0.0243
#>   0.4411  0.2155
#>   0.0950  0.0726
#> [ CPUFloatType{2,3,20} ]
#>