RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9473 -0.2667 -0.4460  0.0868 -0.1724 -0.2901 -0.5792  0.1049 -0.4729
#>   0.4397 -0.4526 -0.8490  0.7701 -0.0529 -0.0730 -0.5488 -0.5972 -0.3877
#>  -0.0481  0.0867 -0.4224 -0.3913 -0.3779  0.8392 -0.5280  0.7610  0.2479
#> 
#> Columns 10 to 18 -0.7653  0.5226  0.0632  0.5856  0.4023  0.1706 -0.5338  0.5791  0.5214
#>  -0.8087 -0.6910  0.5008  0.5406  0.0996 -0.0709  0.5981  0.0032  0.4602
#>  -0.4268 -0.5442  0.1873  0.5735 -0.2471  0.0670 -0.0692  0.6355  0.5781
#> 
#> Columns 19 to 20  0.1226  0.7565
#>  -0.0260  0.0647
#>  -0.2309  0.0376
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3269  0.3766 -0.1388  0.5785  0.2525  0.2820 -0.6601  0.4851  0.6517
#>  -0.1626  0.5564 -0.2114  0.1237  0.5984  0.5517 -0.5289  0.4484  0.2009
#>  -0.3606  0.6690 -0.4163 -0.2292  0.2345  0.0430 -0.3932  0.3179  0.4428
#> 
#> Columns 10 to 18  0.5245 -0.0223 -0.3719  0.6970  0.4135  0.6452 -0.3241  0.4105 -0.3936
#>  -0.3423 -0.2312 -0.0518  0.4660 -0.0935  0.1990 -0.6117  0.3563  0.1493
#>   0.2390  0.6485  0.0607  0.3546  0.2673  0.3080 -0.3760  0.4636  0.3681
#> 
#> Columns 19 to 20 -0.3239  0.2271
#>   0.1981  0.5430
#>  -0.6175  0.1721
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1688 -0.2914 -0.1197  0.4589 -0.0577 -0.4259 -0.5398 -0.2625  0.1214
#>   0.0699 -0.4431 -0.3801  0.0387 -0.2099 -0.1534 -0.4133  0.4116  0.1635
#>  -0.7054  0.1137 -0.4198  0.3141  0.1196  0.2840 -0.9149  0.1046  0.1007
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1176  0.8609 -0.5531 -0.4523 -0.3858 -0.5590 -0.1136 -0.4000  0.1404
#>   0.2853  0.3591 -0.1872  0.2559 -0.2769 -0.4951  0.2873  0.2983 -0.2880
#>   0.2246  0.1977  0.0016 -0.5567 -0.6017  0.0318 -0.1540 -0.1164 -0.2748
#> 
#> Columns 10 to 18  0.7749 -0.4355  0.4041 -0.6540  0.5901  0.6267  0.8868 -0.1121  0.0082
#>  -0.6759 -0.0797 -0.3276  0.6170  0.2070 -0.3228 -0.1796 -0.4801  0.1573
#>   0.1212 -0.4281  0.0421  0.5770 -0.1589  0.3197  0.4962  0.5856  0.1999
#> 
#> Columns 19 to 20 -0.5591 -0.2932
#>  -0.2983 -0.0796
#>  -0.6202  0.0106
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3234 -0.0542 -0.6254 -0.0264 -0.0691  0.1098 -0.5321  0.4875 -0.2775
#>  -0.5667 -0.3431  0.0338  0.4688 -0.3453  0.0196 -0.3739 -0.0944 -0.1847
#>  -0.6149  0.1965 -0.3395  0.1313 -0.2310  0.4687 -0.1945  0.2024  0.2709
#> 
#> Columns 10 to 18 -0.5223 -0.2731  0.2280  0.7393  0.2238  0.5263 -0.1959  0.1848  0.1355
#>  -0.0395  0.1040  0.0880  0.5019  0.0376  0.2721 -0.3948  0.1821 -0.1897
#>  -0.0889 -0.1130  0.2652  0.6497  0.3061  0.4659 -0.4787  0.0235 -0.1336
#> 
#> Columns 19 to 20 -0.4368  0.7379
#>  -0.0571  0.0568
#>  -0.4157  0.1701
#> [ CPUFloatType{2,3,20} ]
#>