RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4360 -0.3336  0.3602 -0.9418  0.7114  0.4844 -0.5811 -0.3604 -0.3655
#>  -0.6302 -0.5428  0.2278 -0.7391  0.6087  0.6547 -0.4779 -0.0270 -0.5866
#>   0.8576 -0.7080  0.6242 -0.6214 -0.5532 -0.6096  0.7104 -0.1501 -0.6846
#> 
#> Columns 10 to 18  0.7264  0.5942  0.1323  0.8222 -0.5565  0.2518  0.4896  0.1973  0.7921
#>  -0.2703 -0.0574 -0.3364  0.1524 -0.4550  0.4722 -0.6677 -0.0652 -0.3504
#>  -0.8536 -0.2273  0.1583 -0.0418  0.5906  0.3778 -0.5238 -0.9590 -0.3130
#> 
#> Columns 19 to 20  0.6045 -0.6609
#>  -0.7669 -0.1975
#>  -0.1791 -0.5472
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1469 -0.4129 -0.0685 -0.3739  0.3070 -0.4193  0.2803 -0.6129 -0.3641
#>  -0.2648 -0.7515 -0.0133 -0.4432  0.1050  0.3585 -0.4634 -0.1361 -0.6881
#>  -0.3385 -0.7463  0.0805 -0.3783  0.5548  0.5783 -0.1299  0.1257 -0.2744
#> 
#> Columns 10 to 18 -0.4605 -0.1305 -0.7239 -0.0994 -0.4257 -0.3066  0.0498 -0.0231  0.0349
#>  -0.4881  0.7363 -0.1995  0.0639 -0.7170  0.4055 -0.0233 -0.1747  0.0868
#>  -0.2487  0.0882 -0.6540  0.0698  0.0450  0.0419  0.3381  0.1422 -0.1735
#> 
#> Columns 19 to 20 -0.8037 -0.4934
#>  -0.0300 -0.5709
#>   0.4288 -0.5839
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1041 -0.7266 -0.0700 -0.4142  0.3359 -0.0900 -0.4664 -0.2672 -0.6199
#>  -0.3366 -0.4746 -0.0577 -0.2341  0.3805 -0.2541  0.1496 -0.7379 -0.5282
#>   0.1548 -0.7415  0.3326 -0.7772  0.2980 -0.1158 -0.1716 -0.4998 -0.4022
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6442 -0.6185  0.7086 -0.4311  0.6832  0.5834 -0.1370 -0.1149 -0.8356
#>  -0.7420 -0.3023  0.5295 -0.4380 -0.4070 -0.6205 -0.3675  0.4230 -0.2029
#>  -0.7652 -0.0638  0.8069 -0.3026 -0.0335 -0.1488  0.1350 -0.7939 -0.2521
#> 
#> Columns 10 to 18 -0.2786 -0.3905  0.7271 -0.2281 -0.4808 -0.2052  0.7602 -0.4605  0.3153
#>  -0.1274  0.0288 -0.0772  0.0986 -0.3559  0.2515 -0.2758  0.6269 -0.4382
#>  -0.3421  0.4637 -0.5389  0.1600  0.5122  0.1416  0.3422 -0.0829  0.3133
#> 
#> Columns 19 to 20 -0.9205  0.2794
#>  -0.4113  0.1102
#>  -0.6111 -0.0605
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1231 -0.7950  0.2350 -0.6020  0.3646  0.5988 -0.6173  0.1156 -0.5667
#>  -0.0122 -0.6416  0.2911 -0.2956  0.4106 -0.0233 -0.4247 -0.5892 -0.0849
#>  -0.2969 -0.4319  0.0712 -0.1268  0.6557  0.5987  0.0328 -0.3051 -0.3720
#> 
#> Columns 10 to 18 -0.4149  0.5632 -0.5008  0.4383 -0.4525 -0.0582 -0.3062 -0.3166 -0.1810
#>  -0.3253  0.0353 -0.3748 -0.1815  0.0297  0.2696 -0.1892  0.3638 -0.0678
#>  -0.1754  0.1960 -0.5509  0.1196 -0.5239 -0.4031  0.1335  0.3921  0.0041
#> 
#> Columns 19 to 20 -0.0706 -0.4365
#>   0.0327 -0.5775
#>  -0.2336 -0.5884
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>