RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4745 -0.1859  0.1777 -0.1858  0.5894  0.2637 -0.6385 -0.7097 -0.7912
#>   0.4220  0.6160 -0.5404  0.7711  0.4409  0.9132 -0.4366  0.4381 -0.0477
#>   0.4044 -0.0396 -0.1762 -0.0959  0.3090  0.5827 -0.2244  0.5444 -0.2442
#> 
#> Columns 10 to 18  0.6514  0.0520 -0.7184 -0.0654  0.4755  0.1592 -0.3316 -0.0359 -0.0747
#>  -0.5912 -0.6414  0.5730 -0.5611  0.3446  0.3079  0.4660 -0.5513 -0.7317
#>  -0.4048 -0.4203  0.0843  0.0599  0.5302 -0.1320  0.7580  0.5875  0.6088
#> 
#> Columns 19 to 20 -0.0469  0.7087
#>   0.5845 -0.0100
#>  -0.6411 -0.8461
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4009  0.2581 -0.0894 -0.1635  0.4117 -0.2862 -0.3065  0.4298 -0.1099
#>  -0.1908  0.7495  0.2457  0.2087  0.5420 -0.3270 -0.6689  0.2622  0.7032
#>   0.3058  0.4054 -0.2991 -0.2056  0.0978  0.0599 -0.5669  0.2577  0.3610
#> 
#> Columns 10 to 18 -0.2357  0.0171  0.1912 -0.0659 -0.2354  0.0344 -0.4921 -0.7601 -0.3682
#>  -0.3424  0.4032 -0.0949 -0.3796  0.7319 -0.5674 -0.1728  0.2599 -0.0181
#>  -0.4019  0.2927  0.2896 -0.6814  0.6901  0.4075  0.0290  0.4181  0.2707
#> 
#> Columns 19 to 20  0.7225  0.2856
#>  -0.2969 -0.0214
#>  -0.6084 -0.0990
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1321 -0.0842  0.0630  0.0963  0.6189  0.4216 -0.1391  0.1832  0.0561
#>   0.2111  0.4532 -0.3053  0.1982  0.5539  0.3073 -0.7156  0.4295 -0.1437
#>   0.1695  0.1270 -0.0628 -0.4363  0.6908  0.2328 -0.4889  0.4444 -0.0808
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1138 -0.2225  0.0216 -0.0199  0.0298  0.0759 -0.4418  0.1248  0.3882
#>  -0.3342  0.0456  0.1421  0.2609  0.3095 -0.1813  0.0032  0.2931  0.0233
#>  -0.2491 -0.0110  0.1530  0.1752 -0.5961  0.4376  0.7177 -0.6412 -0.4838
#> 
#> Columns 10 to 18  0.1736 -0.0666 -0.2277 -0.7187 -0.5891  0.4111 -0.1654 -0.1506 -0.1801
#>  -0.6499 -0.7127  0.6249  0.0295  0.4223  0.6072  0.0140 -0.2525 -0.6280
#>  -0.1540  0.4431 -0.8291 -0.8185 -0.5821 -0.3455 -0.0026 -0.2175 -0.2324
#> 
#> Columns 19 to 20  0.6187 -0.0353
#>  -0.5784  0.3641
#>   0.0805 -0.6822
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2262  0.1554 -0.0319 -0.2820  0.7412  0.6688 -0.3427  0.3909 -0.1387
#>   0.2347  0.1486  0.2258 -0.1337  0.6556 -0.3578 -0.5898  0.3473 -0.4774
#>   0.1207  0.4732 -0.1367  0.1095  0.5994  0.3849 -0.5638 -0.0245  0.3990
#> 
#> Columns 10 to 18  0.1378 -0.0556 -0.1344 -0.0664  0.3107 -0.1199  0.0441 -0.1585 -0.3549
#>  -0.4922  0.1313 -0.1017 -0.1081  0.5549 -0.3796 -0.1600  0.1735 -0.7019
#>  -0.0577  0.1594 -0.1253 -0.2464  0.5878  0.1730 -0.1709 -0.2612  0.0503
#> 
#> Columns 19 to 20  0.4534 -0.3872
#>   0.2702 -0.3147
#>   0.1245 -0.1123
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>