RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7782  0.5319  0.7348 -0.5717 -0.3200  0.2645 -0.4136 -0.4185  0.4689
#>   0.4997  0.5594  0.1728 -0.4232 -0.3960  0.7110  0.4633  0.1456 -0.4034
#>   0.0192  0.1438  0.8690  0.2932  0.7399  0.4316 -0.6027  0.5309 -0.8528
#> 
#> Columns 10 to 18 -0.4602 -0.0206  0.4422 -0.8845  0.8314  0.6283  0.5087 -0.0046 -0.6192
#>  -0.4841  0.0228 -0.5453 -0.0346  0.2057 -0.2009 -0.5343 -0.4393  0.9020
#>  -0.7854 -0.9357  0.1002 -0.8227  0.9330  0.2925 -0.4544 -0.5494 -0.1651
#> 
#> Columns 19 to 20  0.2260  0.1950
#>   0.3726  0.2584
#>   0.2922 -0.4758
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4266  0.1937 -0.3006 -0.6292  0.1187  0.1125 -0.4291  0.7212  0.2531
#>   0.3200 -0.5068  0.3187 -0.2364  0.4137  0.2276 -0.1675  0.5162 -0.0862
#>   0.3342 -0.1007  0.2868 -0.7673  0.5610  0.4554 -0.2307  0.3019  0.5724
#> 
#> Columns 10 to 18 -0.1192 -0.0562  0.1152 -0.3838 -0.2364  0.2255  0.3947  0.0994  0.0006
#>  -0.6363 -0.3362  0.3257  0.3201  0.6479 -0.1272  0.2504 -0.2467 -0.0725
#>   0.5838  0.5401  0.4521 -0.0595 -0.1674  0.4655  0.7744 -0.0487  0.4785
#> 
#> Columns 19 to 20  0.5753  0.0840
#>  -0.0259  0.1098
#>   0.1681  0.2872
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0196  0.0045  0.6745 -0.1057  0.3297  0.6667 -0.3754  0.5296  0.3230
#>   0.5449 -0.1957  0.0307 -0.1658  0.1845  0.2984 -0.5669  0.3841  0.2429
#>  -0.2508  0.2058  0.1293 -0.0970  0.0652  0.1518 -0.8217  0.8132  0.0587
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5953  0.0710 -0.2327  0.1677 -0.6255  0.3973  0.0578 -0.5143 -0.7132
#>   0.7386 -0.2913  0.4116 -0.4018 -0.5387  0.6609  0.1953 -0.1369 -0.4420
#>  -0.1034  0.0109 -0.6470  0.0101  0.0995 -0.0032  0.1650 -0.0833  0.6358
#> 
#> Columns 10 to 18  0.5722 -0.0487  0.2877 -0.2714 -0.6362  0.7201  0.5752 -0.4507  0.4679
#>   0.7528  0.5488  0.1410 -0.7549  0.1960  0.2249  0.1261 -0.1938  0.3264
#>   0.0435 -0.4954  0.0470  0.4881 -0.4157  0.5943 -0.0227  0.4314  0.0214
#> 
#> Columns 19 to 20 -0.6312 -0.2149
#>  -0.5877 -0.2351
#>  -0.0545 -0.2045
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0024 -0.1027  0.2565 -0.0861  0.4345  0.4474 -0.6893  0.6077  0.4612
#>   0.2479 -0.2109  0.0835 -0.1289  0.3989  0.3073 -0.2918  0.4809  0.2586
#>   0.0574  0.1062  0.1841 -0.2528  0.2550  0.3686 -0.5764  0.7056  0.2017
#> 
#> Columns 10 to 18 -0.5450  0.0025  0.3976 -0.2119 -0.7373 -0.1176  0.3024  0.4537 -0.2522
#>  -0.7511  0.1259  0.3960  0.1688 -0.2771 -0.4113  0.0804  0.4265 -0.2836
#>   0.0536 -0.4313  0.2723 -0.3955 -0.2330  0.0913  0.5155  0.0038 -0.0345
#> 
#> Columns 19 to 20  0.0947  0.3794
#>   0.0826  0.2628
#>   0.2290 -0.0622
#> [ CPUFloatType{2,3,20} ]
#>