RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.7246  0.7191 -0.4211 -0.6017  0.2425  0.1399  0.1315  0.5677 -0.7353
#>  -0.4254 -0.6893  0.8209  0.4970 -0.3183  0.5921  0.1227  0.6456  0.8706
#>  -0.1324 -0.0191 -0.5164  0.7798  0.7928 -0.6876  0.3176  0.3528  0.2507
#> 
#> Columns 10 to 18 -0.3179 -0.1779  0.0247 -0.6295 -0.0894  0.2188  0.5280 -0.5378 -0.7696
#>  -0.1188  0.7422 -0.1100 -0.7096  0.3967  0.7176  0.2287  0.8588 -0.6946
#>  -0.7114 -0.1843 -0.2920 -0.1105 -0.6930 -0.0162  0.5158 -0.2727 -0.3791
#> 
#> Columns 19 to 20  0.5064  0.3893
#>   0.7731  0.1411
#>   0.9011 -0.4564
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5834  0.1604 -0.1542  0.2973 -0.2608  0.1410  0.2901  0.3321 -0.2629
#>   0.3684  0.6797  0.3054 -0.0976 -0.1407  0.1415 -0.1872  0.6816  0.1407
#>   0.1641  0.0166  0.6080  0.0693  0.3651 -0.0317  0.6854 -0.0809  0.0629
#> 
#> Columns 10 to 18 -0.6324  0.0883  0.4818  0.4258 -0.1735  0.3889  0.2922 -0.3404 -0.1146
#>  -0.0561  0.0881 -0.3597 -0.7298  0.2647 -0.0439  0.0541 -0.0611 -0.1197
#>  -0.5504  0.3479 -0.0332 -0.5278  0.0802  0.3827  0.5975  0.1315  0.0344
#> 
#> Columns 19 to 20  0.7329  0.1268
#>   0.3169 -0.1918
#>   0.3422  0.2646
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2986  0.1115  0.3631  0.3610  0.1527  0.1883 -0.0031  0.2766 -0.1188
#>   0.5386  0.5961  0.1368  0.3962 -0.3279 -0.0200  0.4906  0.7041 -0.3499
#>   0.1844  0.1831  0.2428  0.2527  0.2918  0.3156  0.2169  0.3023  0.2322
#> 
#> Columns 10 to 18 -0.1882 -0.0353 -0.0591 -0.0419 -0.1580  0.1409  0.1456 -0.3973  0.0819
#>  -0.0943 -0.0153 -0.3372  0.2752  0.1887  0.1042  0.4672 -0.3944 -0.1765
#>  -0.2611  0.1414 -0.0058 -0.5217 -0.2132  0.0886  0.3301 -0.2089 -0.2109
#> 
#> Columns 19 to 20  0.7031  0.1360
#>   0.7788 -0.0849
#>   0.6493  0.4263
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.3729  0.2831  0.5149  0.4924  0.2466 -0.2009  0.3798  0.0027 -0.3332
#>   0.2338  0.2341  0.3775  0.2640 -0.0538  0.6154  0.0863  0.4404 -0.1072
#>   0.3157  0.3227  0.1460  0.0056  0.2435 -0.0715  0.0811  0.3933 -0.0647
#> 
#> Columns 10 to 18 -0.3212 -0.1151 -0.2425  0.2109  0.0021  0.1648  0.1482 -0.3364  0.3006
#>  -0.5774 -0.4006 -0.2035  0.0512  0.2947  0.0246  0.0893 -0.4517  0.1192
#>  -0.4267 -0.1022 -0.3394 -0.3380 -0.1663  0.1222  0.5545 -0.3012 -0.3708
#> 
#> Columns 19 to 20  0.7447  0.2325
#>   0.5818 -0.0213
#>   0.7528 -0.0436
#> 
#> (5,.,.) = 
#>  Columns 1 to 9 -0.2765 -0.1153  0.4778  0.1702  0.3532  0.3552 -0.1878 -0.0359 -0.1656
#>   0.4093  0.1200  0.2154  0.3891  0.0451  0.4656 -0.0473  0.3831 -0.2376
#>   0.4602  0.0493  0.3235  0.1698 -0.2279  0.3162  0.2708  0.2567  0.1761
#> 
#> Columns 10 to 18 -0.5160 -0.3030 -0.0390 -0.5393 -0.0968 -0.2056 -0.3285 -0.0932  0.4249
#>  -0.2810 -0.3762 -0.5115  0.0749  0.1555 -0.2356 -0.1413 -0.6197 -0.0558
#>  -0.3109  0.2101 -0.0194 -0.3808  0.0354  0.0800 -0.0185 -0.2781  0.0745
#> 
#> Columns 19 to 20  0.6785 -0.1114
#>   0.5439 -0.1691
#>   0.5754  0.2619
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.2072 -0.1687  0.6571 -0.3512 -0.4730  0.3558 -0.5527 -0.2153 -0.2802
#>  -0.4398 -0.3190 -0.0698  0.5939  0.5868  0.2241 -0.3288  0.2836  0.2924
#>   0.0959  0.3451  0.2281 -0.6906  0.5485 -0.5712  0.2239  0.1446 -0.4135
#> 
#> Columns 10 to 18  0.4248 -0.3199  0.4081 -0.4106  0.6588 -0.2101  0.4585 -0.4496  0.6113
#>   0.3872 -0.2430  0.2030  0.1245 -0.5706 -0.3161  0.0716 -0.2665 -0.4285
#>  -0.0821 -0.3990 -0.4906  0.6404  0.1344 -0.0179 -0.1249 -0.5386 -0.2647
#> 
#> Columns 19 to 20  0.5184 -0.3124
#>   0.2310  0.1672
#>   0.3990 -0.4330
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2765 -0.1153  0.4778  0.1702  0.3532  0.3552 -0.1878 -0.0359 -0.1656
#>   0.4093  0.1200  0.2154  0.3891  0.0451  0.4656 -0.0473  0.3831 -0.2376
#>   0.4602  0.0493  0.3235  0.1698 -0.2279  0.3162  0.2708  0.2567  0.1761
#> 
#> Columns 10 to 18 -0.5160 -0.3030 -0.0390 -0.5393 -0.0968 -0.2056 -0.3285 -0.0932  0.4249
#>  -0.2810 -0.3762 -0.5115  0.0749  0.1555 -0.2356 -0.1413 -0.6197 -0.0558
#>  -0.3109  0.2101 -0.0194 -0.3808  0.0354  0.0800 -0.0185 -0.2781  0.0745
#> 
#> Columns 19 to 20  0.6785 -0.1114
#>   0.5439 -0.1691
#>   0.5754  0.2619
#> [ CPUFloatType{2,3,20} ]
#>