RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9604  0.0520 -0.0148 -0.7836 -0.9665 -0.6389  0.4733  0.4426  0.3029
#>   0.4839 -0.3740 -0.2179 -0.4375 -0.3793  0.2086  0.3183 -0.2424  0.5102
#>   0.2980 -0.1246  0.6677  0.9236  0.4300  0.3448 -0.3783 -0.4204  0.1375
#> 
#> Columns 10 to 18 -0.2494  0.1558 -0.1704 -0.5940  0.0941 -0.3544  0.8812  0.6898 -0.7727
#>  -0.2140 -0.1887 -0.8498  0.7259  0.2385 -0.5118  0.7284  0.3898  0.0146
#>   0.0627 -0.8335 -0.6561  0.8615  0.6774 -0.1693 -0.6683  0.4381  0.1045
#> 
#> Columns 19 to 20 -0.8894 -0.6862
#>  -0.4569 -0.6539
#>   0.7691 -0.5496
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5873 -0.3320 -0.4655  0.3589 -0.4010 -0.2861 -0.4597  0.2436  0.6869
#>  -0.2507 -0.1012 -0.0551 -0.1653 -0.1614  0.2481  0.2067 -0.0145  0.3754
#>  -0.1342  0.2989 -0.2750  0.4280  0.3267  0.6758 -0.1787 -0.0050  0.3662
#> 
#> Columns 10 to 18 -0.1845  0.2313  0.5076  0.4776  0.2313  0.2015  0.6952  0.7406  0.4356
#>  -0.4856  0.4912  0.0898  0.5270 -0.4754 -0.6332  0.4657  0.2706 -0.1946
#>   0.4522 -0.6439 -0.0934  0.4288 -0.2103 -0.5975 -0.0286  0.1562  0.0766
#> 
#> Columns 19 to 20 -0.2275  0.7689
#>  -0.3299  0.0233
#>   0.1530 -0.5005
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1639 -0.3937 -0.2611 -0.2328 -0.0141  0.0139 -0.5663  0.1080  0.0508
#>  -0.6076  0.4241 -0.5021  0.0412 -0.0740 -0.2421 -0.0608 -0.2278  0.0825
#>  -0.5220  0.1929 -0.6435  0.4275  0.4370 -0.4624 -0.1035 -0.0220  0.4772
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0685 -0.9610  0.5660 -0.1324  0.8317  0.6892 -0.4457  0.7936  0.5950
#>   0.0248  0.3444  0.7890  0.2598  0.2182  0.1376 -0.6932  0.1441  0.4156
#>  -0.2743  0.0240  0.2786 -0.1107 -0.0564  0.4665 -0.2027 -0.3342  0.4904
#> 
#> Columns 10 to 18 -0.1550 -0.1187 -0.4343 -0.4353  0.7200  0.1035 -0.5177 -0.2745 -0.0393
#>   0.2159  0.1206  0.5302  0.2492  0.4894  0.3822  0.1740 -0.1935 -0.4263
#>  -0.4796  0.0734 -0.1243  0.1446 -0.2648 -0.3534 -0.5824 -0.4856  0.1714
#> 
#> Columns 19 to 20 -0.2860 -0.1929
#>  -0.0648  0.3089
#>   0.1201  0.0958
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1250  0.1279 -0.8338  0.0716  0.2197  0.4171 -0.2364 -0.1390 -0.2471
#>  -0.5640 -0.3462 -0.1967  0.4137 -0.0354 -0.3753 -0.4864 -0.1396  0.6406
#>  -0.1880  0.1424 -0.2957 -0.0219  0.2994 -0.0949 -0.2211 -0.3279  0.1196
#> 
#> Columns 10 to 18 -0.3955  0.0132  0.4050 -0.0758 -0.4022 -0.0405 -0.3194 -0.0228 -0.3643
#>  -0.1238 -0.4155  0.0869  0.1142 -0.0009 -0.3521  0.5170  0.4598 -0.0880
#>   0.0791  0.1131 -0.3234  0.1597 -0.4937 -0.0419 -0.2798 -0.3699 -0.0325
#> 
#> Columns 19 to 20  0.5492 -0.1386
#>  -0.2009  0.0388
#>   0.4467 -0.0067
#> [ CPUFloatType{2,3,20} ]
#>