RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0850  0.9316  0.6665  0.0240 -0.3938  0.0230 -0.4130 -0.4006  0.7599
#>   0.6771  0.1505  0.7962 -0.1301 -0.2709  0.0587 -0.8560 -0.4063  0.3657
#>   0.6821  0.0899 -0.2680 -0.0848  0.9679 -0.1169  0.1639  0.5543 -0.8888
#> 
#> Columns 10 to 18  0.5360 -0.6671 -0.7005 -0.7220 -0.8760 -0.2676  0.1002 -0.6831 -0.0220
#>  -0.7372 -0.4812  0.8014  0.3983 -0.7349  0.1359 -0.3635 -0.4071  0.6555
#>  -0.5478  0.7045  0.0077  0.0385  0.0624  0.5424  0.1680  0.7408  0.2209
#> 
#> Columns 19 to 20  0.8228  0.2964
#>   0.7210  0.5597
#>   0.1661  0.2708
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4187  0.1941 -0.0214  0.0829  0.5635 -0.1621  0.1767  0.7022 -0.1960
#>   0.2349 -0.5756  0.1032  0.3000  0.3488  0.4864 -0.4128  0.3264 -0.4373
#>  -0.0968 -0.1150  0.6071  0.1065 -0.3423  0.4876  0.1377 -0.1920  0.4740
#> 
#> Columns 10 to 18 -0.5519  0.0556  0.1040  0.0344 -0.4218  0.6865  0.1236  0.5256 -0.0962
#>  -0.5039  0.2656  0.1393 -0.4363 -0.5013 -0.2668  0.1469  0.4591 -0.4786
#>  -0.1870  0.0399 -0.3070  0.4578 -0.0596 -0.5494  0.8381  0.5723 -0.3508
#> 
#> Columns 19 to 20  0.8022 -0.2764
#>   0.4756  0.0806
#>   0.7203 -0.2334
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1730  0.1278  0.1870  0.0275  0.0925  0.1994  0.0957  0.2087  0.3018
#>   0.2156 -0.1411  0.4181 -0.1840  0.0150  0.2856  0.0708  0.1002  0.5625
#>   0.5665  0.2244  0.5818 -0.0458 -0.3350 -0.2501 -0.4431  0.1464 -0.1805
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6328 -0.2069  0.4042 -0.6809 -0.0327 -0.7630  0.7623 -0.0314 -0.6978
#>   0.0825  0.2147  0.0572 -0.3630  0.5669 -0.2239  0.1057 -0.5619  0.2822
#>  -0.4070  0.0382 -0.3375 -0.5201 -0.7547 -0.8726  0.7777 -0.8260 -0.0959
#> 
#> Columns 10 to 18 -0.1219 -0.6049 -0.0695 -0.4063 -0.3942 -0.0336  0.1864  0.3758  0.2458
#>   0.6062 -0.1832 -0.3139  0.3612 -0.4182  0.1999  0.3858  0.3983  0.3851
#>  -0.2490 -0.4858  0.0320 -0.5870 -0.5178  0.2122  0.4532  0.1370 -0.6275
#> 
#> Columns 19 to 20 -0.3572  0.6884
#>  -0.4728  0.1678
#>  -0.0241  0.1916
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0133 -0.0450 -0.2788  0.1836  0.5438  0.0684  0.2484  0.2174  0.0142
#>   0.1333 -0.2667  0.1874  0.2305  0.2915  0.2595 -0.0629  0.1400  0.1546
#>   0.5105 -0.1352  0.1301 -0.1334  0.3908 -0.4151  0.0125 -0.1494  0.4591
#> 
#> Columns 10 to 18 -0.1187  0.4223 -0.2356  0.1082 -0.0104  0.7285  0.7644  0.1197 -0.0384
#>   0.1300  0.4264 -0.0713  0.1161 -0.3612  0.2876  0.6810  0.4650 -0.3160
#>   0.2683  0.5041 -0.2332  0.1120 -0.0301  0.5933  0.7430 -0.3997 -0.4777
#> 
#> Columns 19 to 20  0.5384 -0.2749
#>   0.3994 -0.1773
#>   0.7592 -0.3269
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>