RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7624  0.3833 -0.5505 -0.4256 -0.3568  0.0439  0.6721  0.4453  0.3432
#>  -0.6697  0.8494  0.2666  0.4317 -0.4872  0.0292  0.0319 -0.1474  0.0567
#>   0.1584 -0.2934  0.2214  0.0887  0.8540  0.8275  0.1381 -0.3271 -0.2993
#> 
#> Columns 10 to 18  0.7319  0.5705 -0.0868  0.5130  0.2355  0.6375 -0.5546  0.7820 -0.4117
#>  -0.5109  0.5297 -0.6229  0.5941  0.1309  0.3686 -0.1543  0.6051 -0.3721
#>  -0.7165 -0.6573  0.4631 -0.8865 -0.5824 -0.2116 -0.4084  0.1655 -0.1030
#> 
#> Columns 19 to 20  0.5212  0.5191
#>   0.8297  0.0955
#>  -0.0775  0.5243
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3479  0.5723  0.1912  0.3120  0.3033  0.6335 -0.2425  0.3460  0.2321
#>  -0.1474 -0.2468 -0.2189  0.3081  0.1968  0.7579  0.0085 -0.3548  0.8370
#>  -0.0200  0.5323 -0.0098 -0.0120  0.2797 -0.2441 -0.5462  0.0525  0.0890
#> 
#> Columns 10 to 18  0.3207 -0.3220 -0.4604  0.0538 -0.2167  0.5538 -0.1622  0.1081 -0.0401
#>   0.0920  0.4460  0.3619  0.0835  0.1433  0.4684  0.1716 -0.0619  0.0585
#>  -0.1982  0.1850  0.5132 -0.3967  0.2792  0.1004 -0.1570  0.0648  0.2364
#> 
#> Columns 19 to 20  0.2589  0.3230
#>  -0.0678  0.1398
#>   0.5640 -0.1684
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2575  0.3533 -0.3969 -0.1134  0.2694  0.1907 -0.2075 -0.0648 -0.0376
#>  -0.1913 -0.4079  0.3460  0.3269  0.1159  0.2786 -0.0801  0.0719  0.4225
#>   0.6467  0.4257 -0.5231  0.0410  0.1613  0.1134  0.0251 -0.2418  0.4326
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6237 -0.2454 -0.7525  0.0341  0.5286  0.2403 -0.7578  0.5067  0.5395
#>  -0.6610  0.0773 -0.4713 -0.0426  0.0247  0.8484  0.0509 -0.1182  0.7203
#>  -0.2902  0.4770 -0.2300 -0.4093 -0.0193  0.6863 -0.4784  0.7480  0.6138
#> 
#> Columns 10 to 18  0.2395 -0.7524  0.0008 -0.6074  0.6384 -0.4535 -0.1093  0.6006  0.0678
#>   0.2747 -0.0505  0.6782 -0.0443 -0.5081  0.3599  0.8102  0.3901 -0.6571
#>   0.5419 -0.3898  0.9104  0.1762  0.0587 -0.1075  0.3220 -0.1853 -0.5460
#> 
#> Columns 19 to 20 -0.8175 -0.1553
#>  -0.3774 -0.4368
#>  -0.3575 -0.8487
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.7579  0.6693 -0.4517  0.1372 -0.1234  0.1639  0.3021 -0.2370  0.5710
#>  -0.0520  0.1814  0.1621  0.5260  0.4655  0.2742 -0.1510 -0.2806  0.3211
#>   0.0632  0.1288  0.2091  0.6454  0.2613  0.5076  0.1855 -0.4057  0.6255
#> 
#> Columns 10 to 18  0.2345  0.3634  0.5888  0.2350 -0.2153 -0.0775 -0.3364 -0.1132 -0.4702
#>  -0.0460 -0.0527  0.0199  0.0954 -0.0942  0.6469 -0.0201 -0.0836 -0.4256
#>  -0.1316 -0.0476  0.1929  0.3840  0.0853 -0.1021 -0.0660 -0.4349 -0.3895
#> 
#> Columns 19 to 20  0.5351  0.2297
#>   0.0989 -0.1717
#>   0.3472 -0.1887
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>