RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8331  0.8708 -0.3752 -0.8902  0.0542  0.7375 -0.5519 -0.3153 -0.7622
#>  -0.1849  0.3201 -0.1675  0.7079 -0.1924  0.3639  0.5003 -0.6943  0.0418
#>   0.2536 -0.8143  0.8220  0.7689 -0.1081 -0.4410  0.5755 -0.6307 -0.9375
#> 
#> Columns 10 to 18 -0.8369  0.6372  0.0550 -0.3407 -0.1550 -0.4206  0.5395  0.5753 -0.3959
#>   0.7820  0.3020 -0.7427 -0.0178 -0.6130 -0.0308 -0.6002  0.4148 -0.3876
#>   0.9661 -0.8899 -0.7765 -0.7429 -0.9667 -0.4755 -0.2153  0.0513 -0.3712
#> 
#> Columns 19 to 20 -0.2986  0.6928
#>   0.0988  0.8242
#>   0.5150  0.1524
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0107  0.1134  0.5237  0.6640  0.5839 -0.6687 -0.0021  0.4348 -0.0206
#>  -0.4064  0.2975 -0.0701 -0.4290 -0.1619 -0.1316 -0.4040 -0.4196 -0.1611
#>  -0.0411  0.7088 -0.0463 -0.4595 -0.3882  0.4632 -0.3133  0.0515 -0.1502
#> 
#> Columns 10 to 18 -0.4921 -0.4783 -0.1145  0.3833  0.2463  0.5075 -0.1901  0.1733 -0.3383
#>   0.0930  0.5856 -0.2827 -0.0408  0.7001  0.5226 -0.4389 -0.2080 -0.2691
#>  -0.2279  0.2344 -0.7864 -0.2265  0.2117  0.5967 -0.0879  0.6062 -0.3551
#> 
#> Columns 19 to 20 -0.2970  0.2733
#>  -0.2861  0.4949
#>  -0.5676  0.6165
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0011  0.1512  0.5013  0.3536  0.1694  0.3060  0.2085 -0.3843  0.3189
#>   0.0742 -0.3464  0.4278  0.5133 -0.0030 -0.0526 -0.4577 -0.2418 -0.1981
#>   0.0465  0.1600  0.1451  0.4666  0.4409 -0.3239 -0.3339 -0.2934  0.2296
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.9008  0.4861 -0.5496 -0.0633  0.5716  0.2870 -0.1298 -0.3077 -0.4160
#>  -0.1656 -0.6574  0.8470 -0.6525 -0.7225 -0.5179  0.4033  0.3850 -0.3059
#>   0.1054 -0.8781 -0.3844 -0.2735  0.6095  0.3437  0.1797 -0.6242 -0.6650
#> 
#> Columns 10 to 18 -0.6009  0.1361 -0.2955  0.0637  0.1947 -0.1518 -0.7030 -0.2326  0.2631
#>  -0.3456 -0.9106  0.6021 -0.3853 -0.5427  0.5115  0.7064  0.6302 -0.1135
#>  -0.1365 -0.3588  0.1877 -0.3737  0.2579  0.6208 -0.7592  0.1803 -0.6048
#> 
#> Columns 19 to 20  0.0466  0.2260
#>  -0.6117 -0.7614
#>   0.2030  0.1114
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2728  0.6638  0.4681 -0.1087  0.1952  0.1117 -0.0983 -0.4727  0.2166
#>  -0.0981  0.0041  0.1865  0.4955 -0.0077 -0.0062 -0.6033 -0.1118 -0.3260
#>  -0.4760  0.4625  0.5225  0.3640  0.2795 -0.1102  0.2751 -0.3882 -0.3341
#> 
#> Columns 10 to 18 -0.3550  0.3192 -0.1699  0.1662 -0.4640  0.2206 -0.3522 -0.3019  0.0900
#>   0.3781 -0.0615 -0.6547  0.0881  0.7336  0.5299 -0.1534  0.1156 -0.0174
#>  -0.0528 -0.0957 -0.4059  0.1088  0.3628  0.4545 -0.5218 -0.0655  0.0965
#> 
#> Columns 19 to 20 -0.5718  0.4177
#>   0.0225  0.4497
#>   0.0001  0.1193
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>