RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7072 -0.0423 -0.3359  0.0305 -0.3877 -0.1484  0.0941  0.3675 -0.2850
#>   0.4093 -0.4974  0.5686 -0.7473  0.1995  0.7289  0.1561 -0.2332  0.2364
#>   0.0775  0.7315 -0.8449 -0.1878 -0.2608 -0.5475  0.6881  0.5273 -0.0539
#> 
#> Columns 10 to 18 -0.5814 -0.1288 -0.5792  0.3296 -0.2180  0.2286  0.5284 -0.2064 -0.4854
#>  -0.0518 -0.1221 -0.5506 -0.3784  0.0650  0.7650  0.1251  0.0103 -0.3688
#>   0.2822  0.0296  0.4435  0.5144 -0.3745  0.8064 -0.6535  0.8146  0.5585
#> 
#> Columns 19 to 20  0.4262 -0.8112
#>  -0.7725 -0.3768
#>  -0.0760  0.0290
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5717 -0.3878 -0.4206  0.0417 -0.0234  0.3393 -0.5488 -0.1898  0.2931
#>  -0.0222 -0.1392 -0.0182 -0.1901  0.5903 -0.1017  0.3342  0.0471  0.4418
#>   0.0327 -0.5450  0.5421 -0.4511  0.3855 -0.3670  0.3942  0.7090 -0.3567
#> 
#> Columns 10 to 18 -0.2794 -0.1553 -0.0719 -0.0605  0.5179  0.1414  0.0253 -0.0379 -0.2751
#>   0.3145 -0.3448 -0.3281 -0.5445  0.2840  0.1569 -0.1531  0.3382 -0.0111
#>   0.6107 -0.3544 -0.0556  0.0129 -0.1527  0.3458 -0.3056  0.4384  0.0613
#> 
#> Columns 19 to 20  0.7411 -0.4337
#>   0.2339 -0.2430
#>  -0.1520  0.3928
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0345 -0.0058  0.1669  0.2045  0.4166  0.2352  0.0100  0.1637  0.2478
#>   0.2426 -0.3925  0.2182  0.0262  0.4046  0.0678 -0.0097 -0.0982 -0.0564
#>   0.3596 -0.6297  0.3764 -0.0646 -0.4342 -0.3463 -0.0896  0.0101 -0.3325
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0173  0.2678 -0.7685  0.3402 -0.1425 -0.6391  0.2945 -0.2789  0.2247
#>   0.1934  0.6915 -0.5806  0.7154  0.1526 -0.6724  0.0100  0.4388 -0.1357
#>  -0.0829 -0.4321 -0.7702  0.0983  0.3529  0.7075 -0.6499  0.6940 -0.2618
#> 
#> Columns 10 to 18 -0.2213  0.2676  0.2106 -0.3214 -0.5299 -0.2602  0.4381 -0.2945 -0.3017
#>  -0.2144 -0.2525 -0.4372 -0.2662 -0.5915 -0.3776  0.0882 -0.4374  0.4405
#>   0.3844 -0.0552 -0.2846  0.5114  0.5288  0.1877  0.2894 -0.5937  0.0406
#> 
#> Columns 19 to 20 -0.4857 -0.4487
#>  -0.0266  0.1902
#>   0.6268  0.7357
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4031 -0.2951 -0.3249  0.3345  0.2861 -0.3874  0.0893  0.4607 -0.2356
#>   0.1962 -0.3260  0.1211 -0.0312  0.3571  0.0032  0.1030  0.4702 -0.0104
#>   0.2663  0.2946  0.1212 -0.6467  0.2585  0.1944  0.5369 -0.2323 -0.0876
#> 
#> Columns 10 to 18  0.4250 -0.2621 -0.0463  0.1163  0.0346  0.2934  0.0444 -0.4660 -0.2421
#>   0.5275 -0.0317 -0.1689  0.5081 -0.0050  0.1417  0.1378 -0.6036 -0.1308
#>   0.3588 -0.5083 -0.6599 -0.3631  0.6986  0.2567  0.1206  0.1732 -0.0488
#> 
#> Columns 19 to 20  0.4111  0.5351
#>   0.2110  0.3478
#>  -0.4441 -0.0529
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>