RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4311  0.5296  0.3149 -0.1306 -0.5858 -0.5505 -0.2886 -0.4924 -0.6509
#>  -0.5467 -0.5526  0.6355  0.1782 -0.1867 -0.5890  0.7933  0.2209 -0.5285
#>  -0.0147  0.3879  0.8657 -0.9568  0.6846 -0.2050 -0.7149  0.1352 -0.4622
#> 
#> Columns 10 to 18 -0.5061  0.1201 -0.5751  0.1634 -0.1615 -0.2589  0.2095 -0.0971 -0.3656
#>   0.8488  0.0643 -0.4574 -0.0621  0.3999  0.8031 -0.3874 -0.9060  0.7537
#>   0.2853  0.9199  0.7993  0.6040 -0.2490  0.6849 -0.2155  0.1307 -0.4776
#> 
#> Columns 19 to 20  0.2729  0.5125
#>  -0.0242 -0.3964
#>   0.1918  0.3541
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2907  0.1290  0.4477  0.1184  0.1564  0.5203  0.4457  0.1821 -0.0195
#>   0.5756 -0.4976 -0.0271  0.6940 -0.4917  0.3298  0.0717 -0.2504 -0.1876
#>  -0.2862 -0.1558  0.1464  0.1279  0.3539  0.4795 -0.1452  0.3771 -0.3305
#> 
#> Columns 10 to 18 -0.1082  0.4456  0.0043 -0.5081  0.4258  0.0302  0.6763  0.0701 -0.2544
#>   0.2877  0.5666 -0.5737 -0.2253  0.0343  0.0381  0.8019  0.2092 -0.0067
#>   0.4307  0.4936  0.1109 -0.3444  0.1801  0.5118 -0.5753  0.2294 -0.1798
#> 
#> Columns 19 to 20  0.0647 -0.2421
#>   0.3593  0.2957
#>  -0.5181  0.6882
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1906 -0.0440 -0.1380 -0.3821  0.1492  0.3256  0.3732 -0.0374 -0.2964
#>   0.3079  0.3622 -0.0652 -0.2360 -0.1449  0.5290  0.3739 -0.0400  0.0349
#>   0.7331  0.1148  0.3201 -0.0382 -0.0332  0.4916  0.0155 -0.2155  0.1802
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2657 -0.5690  0.6129 -0.9131  0.0846 -0.5065 -0.4620 -0.4858  0.1151
#>   0.5381  0.6928  0.7524 -0.4761  0.1738 -0.1452  0.2735 -0.0997 -0.6216
#>   0.0754 -0.5573  0.4463 -0.1849  0.5724  0.2895 -0.0600  0.5340 -0.1680
#> 
#> Columns 10 to 18  0.5452  0.2241  0.5639  0.0976 -0.5340 -0.0081  0.2096 -0.1512  0.6633
#>  -0.5375 -0.2075 -0.0314  0.2228 -0.4425  0.5580  0.0232 -0.0492 -0.5716
#>  -0.4003  0.3747  0.2007 -0.0697 -0.3078  0.2641 -0.6138 -0.4736  0.3793
#> 
#> Columns 19 to 20 -0.1216 -0.8092
#>  -0.1608 -0.2289
#>  -0.3731 -0.7968
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5066  0.3101  0.3642  0.1505 -0.3022  0.0915  0.4671  0.1115  0.0647
#>   0.3290  0.4146 -0.5100 -0.1222 -0.0081  0.5252 -0.2192 -0.2917 -0.1874
#>   0.1872 -0.3739  0.5795 -0.1103 -0.4746 -0.0273 -0.1749  0.0229 -0.3443
#> 
#> Columns 10 to 18  0.0101 -0.0546 -0.4016 -0.4012  0.5792  0.5938  0.2540  0.2925 -0.0525
#>   0.1231  0.2077 -0.4132 -0.4318  0.1388  0.1576 -0.1297  0.3056 -0.1180
#>   0.4189  0.5155 -0.3015 -0.2083  0.2253  0.5809  0.2676 -0.2486  0.2099
#> 
#> Columns 19 to 20 -0.0876 -0.0260
#>  -0.5409  0.3393
#>  -0.4953  0.1491
#> [ CPUFloatType{2,3,20} ]
#>