RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1348 -0.2645  0.0108 -0.9432  0.1192 -0.1802 -0.2924 -0.4557  0.8433
#>  -0.8677  0.3553 -0.7593 -0.2593  0.4894 -0.6538 -0.0356 -0.9014 -0.0361
#>   0.2718  0.6146  0.5401 -0.3216 -0.6124  0.3156 -0.9839  0.8818 -0.0950
#> 
#> Columns 10 to 18 -0.4205  0.3852 -0.1571  0.1934 -0.8591 -0.4699  0.2233  0.2857 -0.5149
#>   0.7354  0.2217 -0.1034 -0.7735 -0.2362 -0.3862 -0.4086  0.5162 -0.8166
#>   0.9045  0.3708 -0.5995  0.6675  0.2113  0.8861  0.5812  0.2060  0.2331
#> 
#> Columns 19 to 20 -0.6067  0.0597
#>  -0.0526  0.5303
#>   0.7150  0.9042
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2970  0.5335 -0.2863 -0.1569 -0.2183  0.5466  0.2433 -0.3429  0.0103
#>  -0.4755 -0.2306 -0.8078  0.2526 -0.5519  0.5934  0.2153  0.3022  0.0939
#>  -0.6551 -0.2415 -0.8062 -0.6969  0.3890 -0.3804 -0.6954  0.2428  0.0224
#> 
#> Columns 10 to 18  0.1500 -0.0750 -0.0558 -0.1052  0.1681 -0.7033  0.4306 -0.3011  0.2566
#>  -0.2964  0.2368  0.1404  0.1292  0.7175 -0.3251  0.4261 -0.7046  0.3850
#>   0.2512 -0.0566  0.0794 -0.4268 -0.4411  0.3947  0.2154 -0.8116  0.4288
#> 
#> Columns 19 to 20  0.1015  0.2575
#>   0.0877 -0.2161
#>  -0.7302  0.4336
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0998  0.2625 -0.0848  0.2833 -0.0807 -0.0413 -0.5749 -0.6130  0.5698
#>  -0.5542 -0.1350 -0.4422 -0.4600 -0.1901 -0.6349 -0.0821 -0.2972  0.3647
#>  -0.4319  0.0972 -0.8181 -0.6692 -0.7847 -0.1085 -0.4059  0.1596  0.0238
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0519  0.0727  0.6430 -0.0321  0.5128  0.3913 -0.4345  0.6277  0.6866
#>   0.1249  0.2013 -0.0989 -0.4404  0.3204 -0.3020 -0.3642 -0.3741 -0.0727
#>   0.7913  0.2150  0.7232 -0.0163  0.0617 -0.3464 -0.7365  0.3678  0.2222
#> 
#> Columns 10 to 18 -0.3707 -0.3470 -0.2802 -0.5181 -0.1037 -0.4060 -0.0867 -0.0445  0.4186
#>   0.5884  0.4118 -0.1410 -0.4473  0.0268 -0.7359  0.1160  0.1280  0.3342
#>   0.1676 -0.6168 -0.2377 -0.6363 -0.1933 -0.5056 -0.4282  0.0506  0.1963
#> 
#> Columns 19 to 20 -0.1929  0.3454
#>   0.0708 -0.0216
#>  -0.4936 -0.1860
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0496  0.2659 -0.4532 -0.3442 -0.5422 -0.3205 -0.6636 -0.2246  0.5585
#>  -0.6329  0.1617 -0.3274 -0.1474 -0.4553 -0.2896 -0.1091 -0.4838 -0.3859
#>  -0.5292 -0.0974 -0.6339  0.0069 -0.2416 -0.0743 -0.4690 -0.1544  0.4358
#> 
#> Columns 10 to 18  0.2658  0.4297 -0.2202  0.1995 -0.5211  0.0202  0.3200 -0.1268 -0.0647
#>  -0.3258 -0.1645  0.4539 -0.1314 -0.3832 -0.2179  0.0359 -0.3843 -0.2327
#>   0.1662  0.3517 -0.2822 -0.0583 -0.0581  0.1440  0.1505 -0.6894  0.0026
#> 
#> Columns 19 to 20  0.1877  0.3144
#>  -0.1398  0.5848
#>   0.0780  0.3154
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>