RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0015  0.9162 -0.0417  0.7983 -0.6227  0.9389 -0.1482 -0.6486  0.0312
#>   0.3928  0.6516  0.2390 -0.7284  0.6743 -0.0762  0.2949  0.3395  0.5973
#>  -0.1706 -0.4036  0.5606 -0.2566  0.0520 -0.5674  0.6898  0.2166  0.1538
#> 
#> Columns 10 to 18 -0.0713  0.5888  0.4886  0.4900 -0.6528  0.0749  0.5047  0.0512 -0.5772
#>   0.8518 -0.3401  0.2433 -0.4467 -0.2006  0.6263  0.4344 -0.7613  0.0766
#>   0.2483 -0.5529  0.1615  0.6602  0.6171 -0.0321 -0.7147  0.3587  0.7646
#> 
#> Columns 19 to 20 -0.6398 -0.8191
#>   0.1224 -0.6410
#>   0.2284 -0.3195
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5488  0.7961  0.2190  0.0170  0.3285  0.5303  0.2912 -0.2789  0.5177
#>  -0.2240 -0.1197  0.3036  0.0949  0.1601 -0.2178 -0.4803  0.2828 -0.0273
#>  -0.1800 -0.3899  0.2384 -0.3313  0.1336  0.2092  0.2237  0.1777  0.4650
#> 
#> Columns 10 to 18  0.4079 -0.7159 -0.7500  0.4295  0.1449  0.0736 -0.5596 -0.0612  0.0591
#>   0.0047  0.0626 -0.2174  0.0853  0.0626  0.0049  0.0226 -0.0203  0.3921
#>  -0.7696  0.1483  0.1632  0.4794  0.4600 -0.2257 -0.0068 -0.5705  0.2343
#> 
#> Columns 19 to 20 -0.0798  0.0842
#>  -0.2987 -0.0085
#>  -0.4121 -0.1157
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2241 -0.2028  0.5651 -0.1197  0.1279 -0.2553 -0.0612 -0.2826 -0.2758
#>  -0.1871  0.0818  0.0644 -0.1170 -0.0929 -0.2205  0.1694  0.1320  0.5519
#>   0.3387  0.1983 -0.3930 -0.3844  0.0602  0.5855 -0.2108  0.2756  0.4721
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0073  0.7203 -0.5184 -0.0165  0.8430 -0.2258  0.3352  0.7735 -0.3388
#>  -0.5505 -0.7433 -0.3873  0.0499  0.4861  0.5929 -0.0012  0.5620 -0.1514
#>   0.5024  0.4107 -0.3683  0.2778  0.6376 -0.0594  0.3647  0.6034 -0.3675
#> 
#> Columns 10 to 18  0.1990  0.1376  0.1317 -0.2677 -0.7964 -0.3154 -0.1641  0.1024 -0.3331
#>   0.2298  0.3776 -0.6066  0.5591 -0.7114 -0.0966 -0.5105 -0.2851  0.0493
#>  -0.0737  0.0953  0.1074  0.0054 -0.1783 -0.5120  0.1257 -0.2489 -0.0887
#> 
#> Columns 19 to 20 -0.2988 -0.5670
#>   0.0512  0.1411
#>   0.2942 -0.0586
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1404 -0.2109  0.5958 -0.2928  0.1561 -0.3847 -0.0069 -0.1456 -0.0914
#>   0.2657 -0.1691  0.1805 -0.0869 -0.1627  0.1710  0.3145  0.3105  0.1607
#>   0.2697  0.2791 -0.1769 -0.2236  0.1879 -0.1170  0.0874  0.1855  0.6315
#> 
#> Columns 10 to 18  0.0620 -0.5091 -0.3479  0.4567  0.0718  0.0081 -0.0858 -0.3860 -0.2644
#>  -0.5801 -0.1624  0.1394  0.4136  0.0839 -0.1295 -0.0918 -0.0134 -0.1284
#>   0.2242 -0.3578 -0.2037 -0.0025 -0.1525  0.1273  0.0673 -0.3804 -0.0363
#> 
#> Columns 19 to 20 -0.3805  0.1193
#>  -0.1946  0.0881
#>  -0.0668 -0.0061
#> [ CPUFloatType{2,3,20} ]
#>