RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3904  0.5864 -0.1080  0.7333  0.0825 -0.5863 -0.0208  0.6301 -0.8250
#>  -0.2503 -0.1159  0.4329  0.6452 -0.5003  0.3032  0.1552  0.1109 -0.5104
#>   0.0668  0.4047  0.9048 -0.0464 -0.8508 -0.4652 -0.8642  0.3382  0.6342
#> 
#> Columns 10 to 18 -0.0828  0.7307 -0.2062 -0.2835 -0.3011  0.3652 -0.6347  0.4595  0.2434
#>   0.2032  0.4018 -0.6755 -0.5884 -0.3998 -0.0115 -0.2767 -0.2674  0.0768
#>   0.9638  0.4718 -0.7320 -0.3713 -0.0546  0.5663  0.0408 -0.5613  0.6567
#> 
#> Columns 19 to 20 -0.7481  0.7639
#>  -0.7171  0.2773
#>  -0.7511 -0.2148
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3686  0.5927  0.2618  0.1862 -0.4037 -0.2560 -0.3705  0.0033 -0.0497
#>  -0.2672  0.2109 -0.0363  0.4347 -0.3288 -0.1487  0.5133  0.1851 -0.1868
#>   0.3031 -0.0427 -0.0969  0.4450 -0.0245  0.2962  0.2111  0.1787  0.2985
#> 
#> Columns 10 to 18  0.6925 -0.1336 -0.6914  0.5275 -0.3157  0.5666 -0.0062 -0.1052  0.1477
#>   0.1013  0.4228 -0.8094  0.2444  0.5342  0.3209 -0.3955 -0.4340  0.4013
#>   0.1498 -0.1468 -0.3811  0.1101 -0.2928  0.6966  0.2853 -0.4776  0.0798
#> 
#> Columns 19 to 20 -0.3215 -0.5022
#>  -0.5687 -0.0395
#>  -0.3531 -0.0895
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4928  0.1155  0.4114  0.5480 -0.5734 -0.2396 -0.4625 -0.1346  0.2302
#>  -0.1619  0.2997  0.2490  0.5097  0.1911  0.1894 -0.3850 -0.4418 -0.0977
#>   0.2146  0.1334  0.6491  0.4555 -0.0397  0.1038 -0.2121 -0.1800 -0.2389
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2148 -0.5456 -0.4285  0.8434 -0.3344 -0.1492 -0.0925 -0.4624 -0.4728
#>  -0.2817 -0.7983 -0.7355  0.6514  0.3569  0.4614 -0.4910  0.3219 -0.4213
#>   0.1624 -0.8302 -0.4563  0.3472  0.2611 -0.7854 -0.8519  0.2182 -0.4583
#> 
#> Columns 10 to 18 -0.2333 -0.2640  0.7098  0.5528  0.8507 -0.0069  0.6041  0.2755  0.8193
#>   0.7686  0.2586  0.7136  0.6522  0.7534 -0.4806 -0.1730  0.2409 -0.0698
#>   0.5979 -0.5243  0.0659  0.8700 -0.6982 -0.8663 -0.3690 -0.5383  0.3060
#> 
#> Columns 19 to 20 -0.2545 -0.7100
#>   0.1000 -0.9085
#>  -0.8569 -0.4108
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0368  0.5400  0.2843  0.2895 -0.3694 -0.3781  0.2722  0.1747  0.0945
#>  -0.3189  0.2687  0.2092  0.4934 -0.1143 -0.3882  0.0375  0.3399  0.1120
#>  -0.5309 -0.4036 -0.1668  0.1072 -0.4910  0.0767  0.3877 -0.1577  0.4435
#> 
#> Columns 10 to 18  0.3952  0.1587  0.2688  0.2241  0.2681  0.5301  0.1382  0.3465  0.2887
#>   0.3276  0.4616  0.1057  0.1844 -0.3689  0.2842  0.2137 -0.0472  0.1211
#>   0.3497  0.2702 -0.4553 -0.1992 -0.1192  0.4548  0.0734  0.2840 -0.3170
#> 
#> Columns 19 to 20 -0.8075 -0.4015
#>  -0.6564 -0.3079
#>  -0.8820 -0.4918
#> [ CPUFloatType{2,3,20} ]
#>