RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8417  0.6320  0.8127 -0.8365 -0.7624  0.8792  0.2247 -0.3913  0.2919
#>   0.9167 -0.2102 -0.1084 -0.0433  0.0667  0.2586 -0.1675  0.3286  0.3082
#>   0.3557  0.5188 -0.8655  0.3515 -0.6424  0.1131 -0.0309  0.8251  0.0387
#> 
#> Columns 10 to 18 -0.1710  0.4068  0.6303 -0.2866 -0.4621 -0.7381  0.0185  0.3059 -0.4353
#>   0.1661 -0.3345 -0.2006 -0.5137 -0.2334 -0.5223 -0.4115 -0.5138 -0.0187
#>   0.7355  0.1327  0.1186  0.4355  0.3332 -0.5758  0.6070  0.3498  0.6065
#> 
#> Columns 19 to 20  0.7948 -0.0059
#>  -0.0741  0.8996
#>  -0.1150  0.0161
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1360  0.2661  0.5098 -0.6274 -0.6513  0.6900 -0.4866  0.2355  0.4855
#>   0.4660  0.2348  0.0276 -0.0093 -0.7625  0.2427 -0.1581  0.2072 -0.0361
#>  -0.4357  0.3370  0.4104 -0.1671  0.3094 -0.1990 -0.6746  0.3172 -0.3072
#> 
#> Columns 10 to 18 -0.3869 -0.3786  0.5454 -0.0451 -0.6156 -0.5498  0.2399 -0.5660  0.0252
#>   0.4022 -0.3421 -0.3334  0.0325  0.1191 -0.6772  0.3926 -0.4016 -0.5073
#>   0.0350  0.5700 -0.3946 -0.1778 -0.6786 -0.6385 -0.0148  0.2420  0.4393
#> 
#> Columns 19 to 20  0.3748  0.3468
#>  -0.6003  0.5122
#>  -0.3025  0.0557
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3134  0.5631  0.6898 -0.2283 -0.4698  0.4217 -0.5321 -0.1301  0.3907
#>  -0.0394  0.0606 -0.1395  0.1531 -0.1769  0.1284 -0.5015  0.2751  0.2407
#>   0.0243  0.3639 -0.0567 -0.3162 -0.4171  0.5540 -0.0379  0.3259  0.0446
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5485  0.1664 -0.4617  0.5259 -0.2581 -0.2082  0.3157  0.1904  0.3121
#>   0.2825 -0.2631  0.6352 -0.5614 -0.1890  0.5416 -0.2519 -0.0599 -0.0894
#>   0.7681 -0.7752 -0.0391 -0.3109  0.8160  0.9703 -0.1940 -0.6318 -0.4213
#> 
#> Columns 10 to 18 -0.5207 -0.0410 -0.1370 -0.3222 -0.4237  0.4606  0.2688  0.2988 -0.6468
#>   0.1550  0.5402  0.0382 -0.1132  0.0319 -0.2513 -0.2019  0.1365 -0.5608
#>  -0.0419 -0.0013 -0.3128  0.7316 -0.7637 -0.6114 -0.5781 -0.6778  0.3666
#> 
#> Columns 19 to 20  0.5374 -0.4053
#>  -0.2967 -0.2439
#>   0.1660  0.8376
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1203  0.3545 -0.0787 -0.2351 -0.4267  0.3988 -0.4704  0.4580  0.0796
#>  -0.3352  0.3970  0.2915  0.0190 -0.5117  0.4169 -0.3106  0.0397  0.4282
#>   0.0600  0.1339 -0.6994  0.1600 -0.3959  0.2582 -0.4134 -0.0647  0.1546
#> 
#> Columns 10 to 18 -0.0515 -0.0163 -0.4301 -0.4616  0.0887 -0.4738 -0.0561 -0.3388  0.0014
#>   0.4956  0.1846 -0.4260 -0.2410  0.0065 -0.5050  0.0664  0.1428 -0.2135
#>   0.5723 -0.2226 -0.5048 -0.0372  0.2102 -0.8400  0.0790  0.0219 -0.3010
#> 
#> Columns 19 to 20 -0.1884  0.3156
#>  -0.3896  0.3434
#>  -0.6602  0.2573
#> [ CPUFloatType{2,3,20} ]
#>