RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4025  0.2832 -0.1093  0.7007 -0.3272  0.3468 -0.4691 -0.0449 -0.1301
#>  -0.0302  0.1200  0.2236  0.7337 -0.2362  0.1393  0.6474  0.5917 -0.5349
#>  -0.7875  0.9512 -0.8110 -0.1720 -0.0541 -0.1175 -0.1688 -0.1910 -0.5895
#> 
#> Columns 10 to 18 -0.5194 -0.0386 -0.1606  0.2056 -0.6080  0.1229  0.3747  0.5826 -0.4825
#>   0.3818 -0.2133 -0.8372  0.3213 -0.8373  0.5141  0.4406  0.7651  0.6132
#>   0.5061  0.7146 -0.7992 -0.3386  0.3083  0.3443 -0.0641 -0.5420  0.0966
#> 
#> Columns 19 to 20  0.5279  0.5084
#>   0.0814  0.4112
#>   0.4476 -0.1579
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1703  0.1813 -0.2053  0.1949 -0.2872 -0.0114  0.4108  0.2679 -0.0644
#>   0.1976 -0.2879 -0.0685  0.2624 -0.3981 -0.1907  0.9044  0.1728 -0.0475
#>  -0.2158 -0.1660 -0.0582 -0.2434 -0.4189  0.2097 -0.1388 -0.3574  0.2236
#> 
#> Columns 10 to 18 -0.5660  0.1226 -0.0996 -0.0474 -0.0415  0.4031 -0.0515 -0.4335 -0.7341
#>  -0.7565 -0.3238  0.7757  0.0090 -0.7105  0.3446 -0.2808 -0.3967 -0.8974
#>  -0.3417  0.1398  0.1558  0.2990  0.6693  0.2918  0.3616 -0.2880  0.0087
#> 
#> Columns 19 to 20  0.7246  0.6221
#>  -0.0993  0.5000
#>   0.2763  0.7029
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1566  0.4728 -0.5169 -0.4157  0.1675 -0.1019  0.0937  0.0327  0.2053
#>   0.3098  0.0459 -0.2544  0.4648 -0.0640 -0.0783  0.4764  0.3026 -0.0667
#>  -0.1202  0.2829 -0.2364  0.3866 -0.2885 -0.4073  0.6056 -0.1033 -0.1529
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1486  0.3662  0.2833  0.6212  0.0163  0.2278 -0.1412  0.5360  0.5450
#>   0.1535  0.2647 -0.1121  0.6081 -0.1246  0.6274 -0.1389  0.3522  0.6061
#>   0.4360 -0.0084  0.2012  0.2025 -0.2092  0.6610  0.1269  0.0320  0.6364
#> 
#> Columns 10 to 18 -0.1129 -0.5326  0.7234 -0.1662 -0.4000  0.5884  0.3398 -0.5452 -0.3733
#>  -0.4531 -0.6434  0.7488  0.2410 -0.1002  0.6127  0.2806 -0.1553 -0.4591
#>  -0.3184 -0.5566  0.1006  0.2449  0.0338  0.5688 -0.5439 -0.1002 -0.5197
#> 
#> Columns 19 to 20 -0.2576 -0.1050
#>  -0.0863 -0.5123
#>   0.1224 -0.7942
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0926  0.3530  0.0152 -0.2692 -0.2381  0.2692  0.3080  0.0586 -0.0836
#>   0.3105  0.3593 -0.4029  0.0912 -0.1038  0.1652  0.0869 -0.0274  0.0137
#>   0.0273  0.1968 -0.4026  0.0863 -0.0345 -0.0553  0.2797  0.1455  0.0857
#> 
#> Columns 10 to 18 -0.3206 -0.1969 -0.5792  0.2023 -0.1874  0.7146  0.1860  0.0457 -0.2404
#>  -0.2324  0.2833 -0.5958  0.2801 -0.2779  0.5254  0.4522 -0.0175 -0.1906
#>  -0.6058  0.2552 -0.1538  0.2076 -0.3406  0.4600  0.1940  0.1073 -0.3070
#> 
#> Columns 19 to 20  0.1458  0.6750
#>   0.3077  0.6497
#>   0.3251  0.3762
#> [ CPUFloatType{2,3,20} ]
#>