RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1891  0.3123 -0.1729 -0.0464 -0.1079 -0.7941 -0.3120  0.5002  0.8378
#>   0.5658 -0.2335 -0.3429 -0.4220  0.4501 -0.4014 -0.0083  0.3826  0.6796
#>   0.5734  0.5236 -0.1495 -0.2132 -0.4914  0.3432 -0.1753  0.6243  0.2816
#> 
#> Columns 10 to 18 -0.8290 -0.7493 -0.2780 -0.9012  0.0891 -0.2558 -0.9026 -0.5557 -0.1494
#>   0.3049 -0.0056  0.3385 -0.0585  0.3445  0.4468  0.4003 -0.5512 -0.4063
#>   0.4301  0.0718  0.8648 -0.2380  0.7713  0.0807 -0.9512 -0.1613 -0.0582
#> 
#> Columns 19 to 20 -0.7063  0.0569
#>  -0.3851  0.2512
#>  -0.4528 -0.9722
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2553  0.5669  0.0506  0.3521 -0.2697  0.1797 -0.4574  0.5414  0.3183
#>   0.0026  0.3344  0.0002 -0.5024 -0.6857  0.0029 -0.3269  0.3404 -0.0489
#>  -0.4406  0.3511 -0.4173  0.2055 -0.4774 -0.4548 -0.4852 -0.3225 -0.2094
#> 
#> Columns 10 to 18 -0.5313 -0.7413 -0.2658 -0.1741 -0.7609 -0.7926 -0.2399 -0.4176 -0.6271
#>  -0.0866 -0.3281 -0.3720 -0.0723 -0.7255 -0.0736 -0.4003 -0.3500 -0.3802
#>  -0.4255 -0.2399 -0.3735 -0.1807 -0.2359  0.1343 -0.5129  0.2069 -0.4538
#> 
#> Columns 19 to 20  0.0543 -0.2404
#>  -0.1359  0.0337
#>  -0.2477 -0.4138
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0757 -0.0451 -0.0238  0.4343 -0.2952  0.4134 -0.2923  0.4892 -0.2122
#>  -0.2827  0.1432 -0.0998  0.1729 -0.3761 -0.3209  0.0717 -0.0608  0.3983
#>   0.1558  0.2378  0.0736  0.0843  0.0547 -0.2362 -0.3033  0.5927 -0.4931
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3385  0.3186 -0.6749  0.2445 -0.6351  0.2549 -0.4597  0.5025  0.4376
#>   0.6764 -0.2268 -0.0935  0.5914 -0.1501  0.2355 -0.2385  0.1549  0.7608
#>  -0.3085 -0.1507  0.4442 -0.4953  0.5237  0.8044 -0.0368 -0.2948 -0.3527
#> 
#> Columns 10 to 18 -0.4212 -0.0335  0.0586  0.0452 -0.3084 -0.4034 -0.0174  0.4906  0.6527
#>   0.3227 -0.1629  0.1798 -0.6596 -0.3875  0.0936  0.6475  0.5238  0.1993
#>   0.6698  0.6091  0.2771  0.4696  0.6942 -0.2183 -0.1864  0.0342  0.2170
#> 
#> Columns 19 to 20  0.3407 -0.1991
#>  -0.1929 -0.1445
#>   0.6291 -0.4260
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1559  0.2648 -0.2446  0.2236 -0.2599 -0.1735 -0.1548  0.2232  0.2214
#>  -0.3308  0.4182  0.1372 -0.1675 -0.3680 -0.2366 -0.1365  0.0954 -0.2151
#>  -0.4340  0.2130 -0.0350  0.4044  0.1344 -0.4170 -0.1355  0.1773  0.2589
#> 
#> Columns 10 to 18 -0.8048  0.0275 -0.0772 -0.4798 -0.2601 -0.3788 -0.3602 -0.2623 -0.6086
#>  -0.0932 -0.1432 -0.3111 -0.5687 -0.5923 -0.5849 -0.4126  0.0372 -0.5598
#>  -0.6138 -0.3209 -0.4017 -0.2835  0.0188 -0.6677 -0.0146  0.2125 -0.2032
#> 
#> Columns 19 to 20  0.0621 -0.2375
#>  -0.2512 -0.6129
#>  -0.3408 -0.0707
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>