RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1035  0.2979 -0.5623 -0.0268  0.1926  0.8460  0.5633 -0.6104  0.0111
#>  -0.4036  0.4717 -0.0332 -0.5995  0.2594  0.4279  0.3592 -0.3793  0.6582
#>   0.1565 -0.0632 -0.8165 -0.5542 -0.2630  0.6035  0.4002 -0.4312  0.0765
#> 
#> Columns 10 to 18 -0.5137 -0.3480 -0.8822  0.5480 -0.5982 -0.7332 -0.7373  0.4018  0.9303
#>   0.1302  0.0414  0.1806 -0.1457 -0.3828 -0.5798  0.4446 -0.5579 -0.3809
#>  -0.4408 -0.4316  0.6619  0.4313 -0.6245 -0.4288  0.4377  0.2082  0.7106
#> 
#> Columns 19 to 20  0.4557 -0.4437
#>   0.7961  0.6009
#>   0.7249 -0.2701
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.8726  0.0632  0.6224 -0.2589  0.4241  0.0304  0.3065  0.2633  0.2167
#>  -0.1345 -0.2139  0.7395 -0.5031  0.3221  0.5136  0.2296 -0.0579  0.5250
#>  -0.5666  0.0633  0.1131 -0.7608  0.4401  0.1318 -0.2808  0.2995  0.3235
#> 
#> Columns 10 to 18 -0.1956 -0.1463  0.2553 -0.3513  0.2660 -0.0998 -0.0771  0.6961 -0.3089
#>   0.0944  0.1706  0.2305  0.0064  0.1027 -0.4975  0.1611 -0.1878 -0.8035
#>  -0.3268  0.5765 -0.2443 -0.6679 -0.3385  0.3588  0.0567  0.4062 -0.4956
#> 
#> Columns 19 to 20  0.7823 -0.0508
#>   0.4624  0.4640
#>   0.6926 -0.2247
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0536 -0.2939 -0.1325 -0.3981  0.0634  0.2808  0.3764 -0.4848  0.3873
#>  -0.0210 -0.0490  0.5294  0.0035  0.4337  0.5535 -0.0185 -0.2572  0.1405
#>  -0.1825 -0.2175  0.2709 -0.5917  0.1376  0.6553  0.1891  0.0432  0.7075
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5885 -0.0884 -0.6054 -0.1822 -0.1363  0.4368 -0.2581 -0.4853 -0.0094
#>  -0.2506  0.0249 -0.1095 -0.5437 -0.4266 -0.4811 -0.1616 -0.5104 -0.0991
#>  -0.7838 -0.2916 -0.3836 -0.1859  0.2656  0.0203  0.3319 -0.3334  0.3284
#> 
#> Columns 10 to 18 -0.3720 -0.2464 -0.3911  0.5873 -0.0010 -0.1817  0.6434 -0.3620  0.3760
#>  -0.1557 -0.1925  0.2573  0.8603  0.4335 -0.8068 -0.0969 -0.5925  0.0627
#>   0.2890 -0.3164 -0.3398  0.7997 -0.1066 -0.3857 -0.1513 -0.0631  0.2764
#> 
#> Columns 19 to 20 -0.3479  0.5097
#>  -0.1676  0.5373
#>  -0.2065  0.3439
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1516 -0.2848  0.3302  0.2207  0.2614  0.1733  0.5118 -0.0421  0.5876
#>  -0.2585  0.1359  0.3454  0.0763  0.5046  0.3976  0.1786 -0.0897  0.5162
#>   0.0326  0.0645  0.3590 -0.0134  0.0060  0.3882  0.5174 -0.1006  0.5195
#> 
#> Columns 10 to 18  0.0656 -0.0595 -0.4293 -0.3619 -0.1684 -0.7218 -0.3501 -0.1897 -0.5085
#>   0.4382  0.1573 -0.0771 -0.0905 -0.5347 -0.6541 -0.2231 -0.0139 -0.6087
#>   0.3257  0.0722  0.1208 -0.3078 -0.1031 -0.5629 -0.1704 -0.0103 -0.5270
#> 
#> Columns 19 to 20  0.4837  0.2642
#>   0.5831 -0.1117
#>   0.1823  0.2366
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>