RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5015 -0.8655 -0.1346 -0.5825 -0.5307 -0.2667 -0.1694 -0.3586  0.6000
#>   0.7778  0.1948 -0.7188 -0.6047 -0.4096  0.7872  0.7703  0.3379 -0.7165
#>  -0.4662 -0.5087 -0.8035  0.0780 -0.4608  0.8055 -0.7285 -0.5198 -0.1308
#> 
#> Columns 10 to 18  0.2196 -0.3497 -0.5993 -0.1781 -0.4387  0.5487  0.1079  0.5940  0.5555
#>  -0.6893  0.4110  0.4523 -0.8691  0.6230 -0.2745 -0.4771 -0.8694  0.4475
#>  -0.6248 -0.7407 -0.6047 -0.1266 -0.6720 -0.0290 -0.7540  0.7119 -0.0188
#> 
#> Columns 19 to 20 -0.6028  0.7887
#>   0.5445  0.4284
#>  -0.5494 -0.1164
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1885 -0.4748  0.1378  0.4451  0.1102 -0.7297  0.4122 -0.2464 -0.0891
#>  -0.3837  0.2143  0.0595  0.0800 -0.0347  0.0521 -0.1277 -0.3482 -0.0814
#>   0.2367 -0.8106  0.0775  0.2556 -0.1536 -0.6828  0.1590 -0.5243 -0.4363
#> 
#> Columns 10 to 18 -0.1444 -0.4896 -0.5397  0.2576 -0.0202 -0.4395  0.2551 -0.2987 -0.4815
#>   0.6222  0.0909 -0.0891 -0.4143 -0.2222 -0.3590 -0.3059 -0.3503 -0.1824
#>   0.0171 -0.4676 -0.4293  0.0827  0.0120  0.1630  0.3904 -0.3085 -0.0811
#> 
#> Columns 19 to 20 -0.7111 -0.0137
#>   0.0718  0.5326
#>  -0.2627  0.1185
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1687 -0.3992 -0.2974  0.2327  0.0576  0.3158  0.4605 -0.1755 -0.2913
#>  -0.6034 -0.1459 -0.1107  0.0652  0.3702 -0.2795 -0.0803 -0.1940  0.0453
#>  -0.0730 -0.4314 -0.6447 -0.1549  0.1909 -0.1632 -0.0172 -0.2158 -0.4668
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7289 -0.4372  0.2023  0.3778 -0.1850 -0.2262 -0.1424 -0.0884 -0.2213
#>   0.3098  0.5204  0.3027  0.1797  0.4697 -0.2861  0.2886 -0.4130 -0.2628
#>   0.8397 -0.0927  0.7872  0.2607 -0.6103  0.6772  0.4531  0.0361 -0.6431
#> 
#> Columns 10 to 18 -0.3057  0.1471  0.1335 -0.7714  0.3298 -0.7518 -0.5376 -0.6403  0.5147
#>  -0.1885 -0.2819 -0.2960 -0.4347  0.5074 -0.6487  0.2603 -0.4934  0.0564
#>   0.2141  0.3357 -0.2482 -0.8348  0.7493 -0.0009 -0.1869 -0.0164  0.8828
#> 
#> Columns 19 to 20 -0.3203  0.2200
#>  -0.0090 -0.2065
#>  -0.3438  0.3028
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2150 -0.2328  0.2008  0.2669  0.3762 -0.6368  0.1677 -0.5483 -0.1290
#>  -0.2269 -0.5137  0.1487 -0.1026  0.2617 -0.2460 -0.0151 -0.3230  0.1410
#>  -0.4970 -0.3252 -0.0358  0.0322  0.2818 -0.3712 -0.0164 -0.2167 -0.3963
#> 
#> Columns 10 to 18  0.4475 -0.0809 -0.2403  0.3038 -0.2289  0.1465  0.3013 -0.2572 -0.5162
#>   0.0373 -0.1353 -0.4668  0.1172 -0.3058  0.3749  0.3372 -0.3739 -0.1187
#>  -0.1704  0.2754 -0.0059 -0.2556 -0.4256  0.2851 -0.0581 -0.6618  0.0639
#> 
#> Columns 19 to 20 -0.4458 -0.1363
#>  -0.2664 -0.0365
#>  -0.7502  0.1406
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>