RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3336  0.3137 -0.3882  0.8244  0.6244 -0.2335 -0.2082  0.2462  0.9776
#>  -0.8210  0.3943 -0.7146  0.5188 -0.1128 -0.2533  0.4458 -0.4529 -0.1527
#>   0.4832  0.6153 -0.9335  0.3737 -0.5289  0.8984  0.0682  0.2316 -0.7261
#> 
#> Columns 10 to 18 -0.9675 -0.0328 -0.0964 -0.3656  0.7335 -0.1973 -0.2491  0.8636 -0.2143
#>   0.2651 -0.9319  0.3346 -0.0782 -0.1500  0.5391 -0.5840  0.0143  0.5619
#>  -0.8071 -0.0875 -0.3440 -0.8819 -0.0815  0.6680  0.4994 -0.0653 -0.1433
#> 
#> Columns 19 to 20 -0.9544  0.2110
#>   0.6469 -0.4114
#>   0.3721  0.4775
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1472  0.5485  0.5767  0.3720  0.2194  0.3845  0.8034 -0.1792  0.6207
#>   0.2464  0.7195  0.3033 -0.3540  0.3318  0.5099  0.6761  0.1959 -0.3784
#>  -0.6932  0.5129  0.0861  0.2045 -0.1252 -0.2635  0.3354  0.4809  0.4623
#> 
#> Columns 10 to 18  0.0436 -0.1974 -0.2532 -0.3041  0.4451  0.3170  0.5156  0.3381 -0.2547
#>  -0.3978  0.2574 -0.5226 -0.2526  0.5796  0.5603 -0.6853  0.3805  0.4319
#>  -0.4201 -0.2705 -0.4354 -0.3778  0.4697  0.8768 -0.5451 -0.4174  0.7014
#> 
#> Columns 19 to 20 -0.4892 -0.2230
#>  -0.8208  0.5719
#>   0.1642  0.7297
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0781  0.0981  0.1369  0.3635 -0.4415 -0.1629  0.5275  0.5663  0.1497
#>  -0.3982  0.4130  0.4209  0.3082 -0.1035 -0.1196  0.6042  0.4751  0.4273
#>  -0.5157  0.6813  0.0784  0.0112 -0.2023  0.2452  0.6213  0.3453  0.5171
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3933 -0.8327 -0.3588  0.6325 -0.4569 -0.1220 -0.1983  0.4915  0.4082
#>  -0.4521  0.4506 -0.4881 -0.0159 -0.7637  0.2166 -0.3646  0.4114  0.2295
#>  -0.4860 -0.1744  0.3086  0.0882 -0.3150  0.1157  0.3554  0.2289  0.0851
#> 
#> Columns 10 to 18  0.7352 -0.2759 -0.1255 -0.2954  0.3478  0.2921 -0.2931 -0.3125  0.1418
#>   0.2818 -0.7425  0.5003 -0.4370 -0.7082 -0.3291 -0.8331  0.2152  0.4588
#>  -0.2337 -0.4312  0.3279 -0.1049 -0.4084  0.5320 -0.3869 -0.5607 -0.3119
#> 
#> Columns 19 to 20  0.4704  0.2621
#>   0.4567 -0.1410
#>  -0.0979  0.4418
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3281  0.6494 -0.0307 -0.2059 -0.0694  0.3355  0.4813  0.3208  0.4743
#>  -0.5245  0.4911  0.0778 -0.2714 -0.3297  0.0531  0.5090  0.3446  0.0983
#>  -0.5390  0.1878  0.2275  0.0553 -0.0866 -0.1266  0.2277  0.2884  0.2975
#> 
#> Columns 10 to 18  0.0161 -0.4657 -0.3918 -0.3591  0.4576  0.7432 -0.4906  0.3307  0.5179
#>   0.2850 -0.6221 -0.1520 -0.1813  0.3859  0.5916 -0.0236 -0.5834  0.3414
#>   0.3165 -0.2331 -0.1402 -0.1653  0.1859  0.7191 -0.3867 -0.4680  0.3119
#> 
#> Columns 19 to 20 -0.3357  0.2644
#>   0.2070  0.6671
#>   0.3071  0.7345
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>