RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0075  0.3608  0.9745 -0.7994  0.0213  0.6418 -0.0465 -0.5730  0.0859
#>   0.7411 -0.0300 -0.5046 -0.2298 -0.1052  0.2847  0.2156 -0.4704  0.4625
#>   0.5715  0.2190  0.5242 -0.0433 -0.7180 -0.2817  0.0901  0.7677  0.0516
#> 
#> Columns 10 to 18  0.0321 -0.4190 -0.7656  0.0949  0.3873  0.4088  0.8884  0.3756  0.0715
#>   0.1913 -0.1632  0.5347 -0.7682  0.0604  0.8135  0.1408  0.5045  0.6496
#>  -0.7980  0.6805  0.5800  0.0156 -0.0981  0.0224 -0.3077  0.6260 -0.4118
#> 
#> Columns 19 to 20  0.7635 -0.7075
#>   0.0992  0.8280
#>   0.2242  0.8935
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5821 -0.1155 -0.6377 -0.4696 -0.2034  0.5173 -0.0529  0.2337 -0.3313
#>   0.2746  0.6776  0.2533 -0.3905 -0.5341  0.1038 -0.1399  0.0856  0.4141
#>  -0.3466  0.3655 -0.2861  0.7157 -0.4919  0.1804  0.3473  0.6183  0.1617
#> 
#> Columns 10 to 18  0.0920  0.6357 -0.4930 -0.5159  0.2583  0.0028 -0.5286  0.2297 -0.1234
#>  -0.0308 -0.0353 -0.4207 -0.3874  0.0298 -0.0908 -0.3838 -0.1699 -0.3135
#>  -0.5778 -0.6908 -0.3170  0.5251  0.3802  0.3687  0.0028  0.2958 -0.1847
#> 
#> Columns 19 to 20 -0.2165  0.7492
#>   0.4746  0.2030
#>   0.4864 -0.5616
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3982 -0.2176  0.1583  0.0316 -0.3675 -0.1042  0.0213  0.6723 -0.3467
#>  -0.2010 -0.0098 -0.2838  0.2788  0.1276  0.1773  0.1636  0.4653 -0.2112
#>  -0.5831 -0.0001  0.4959  0.5677 -0.4400  0.6863 -0.2105  0.3068  0.0411
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6351 -0.2742 -0.2873 -0.4774 -0.0899 -0.9085  0.2272  0.0445 -0.7878
#>   0.4684 -0.3372  0.3725 -0.3952  0.2522 -0.8376  0.2985  0.4223 -0.5008
#>  -0.1361 -0.0561 -0.3615 -0.3642 -0.2196 -0.2823  0.1861  0.1895 -0.7360
#> 
#> Columns 10 to 18  0.6438  0.7497  0.4444 -0.7210 -0.2315  0.3494 -0.7797  0.4840 -0.3904
#>   0.1894 -0.1698  0.8397 -0.7178  0.1831  0.7350  0.0618  0.5419 -0.0211
#>   0.2668  0.0333  0.5493  0.1146 -0.1670  0.1400  0.2877 -0.1214  0.3850
#> 
#> Columns 19 to 20 -0.7175 -0.6666
#>  -0.8077 -0.5089
#>   0.2913 -0.3572
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5911 -0.5202  0.0300  0.5433 -0.1910  0.0190 -0.1776  0.0324 -0.2625
#>  -0.5488 -0.7283 -0.5275  0.1965  0.4111  0.1782 -0.0127  0.5359 -0.2984
#>  -0.3862  0.1128 -0.4107  0.3533 -0.0796  0.4265  0.1144  0.2119  0.2502
#> 
#> Columns 10 to 18 -0.6068  0.5390 -0.5441  0.1082  0.2025 -0.4205  0.3524  0.1857 -0.7367
#>  -0.3447  0.8705  0.3715  0.4594  0.3721 -0.2077  0.1990  0.5650 -0.5134
#>  -0.0076  0.0613 -0.2664 -0.2412  0.4267 -0.1038 -0.0508  0.3547 -0.4317
#> 
#> Columns 19 to 20 -0.5439 -0.6900
#>  -0.0556 -0.1967
#>   0.0421 -0.0380
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>