RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.5582  0.7278  0.6211  0.5719  0.0207  0.4285  0.2903 -0.3532 -0.3731
#>   0.0002  0.4225  0.4750 -0.1690 -0.1890 -0.1460  0.6663  0.8219  0.6393
#>   0.9695  0.3409  0.1173  0.1951 -0.6552  0.0856  0.9560  0.6676 -0.4903
#> 
#> Columns 10 to 18 -0.3274 -0.8149 -0.5726  0.1831  0.7026 -0.0821 -0.4732 -0.1968  0.5242
#>   0.4931  0.0932 -0.7887  0.1924  0.1613  0.5192  0.6546 -0.4496 -0.8238
#>   0.3136 -0.0841 -0.8789 -0.0527 -0.7001 -0.4266 -0.0299 -0.0255 -0.8521
#> 
#> Columns 19 to 20  0.5296  0.2296
#>   0.6476  0.2801
#>   0.7119  0.4396
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1452  0.2876  0.1956  0.2453 -0.5200  0.5955  0.1882  0.6493  0.0866
#>   0.0836  0.6273  0.1235  0.4289 -0.3307  0.0679 -0.2138  0.5808 -0.0271
#>   0.6403  0.4251  0.3893 -0.3101 -0.5000  0.4366  0.3927  0.7424 -0.2421
#> 
#> Columns 10 to 18  0.0028 -0.3509  0.0520 -0.0923  0.1564 -0.3599 -0.3525 -0.1906 -0.3348
#>  -0.4195 -0.3192 -0.3126  0.3521  0.3562  0.0675  0.5907 -0.2607  0.3317
#>  -0.4146 -0.3615 -0.1100  0.2514  0.4553 -0.1491  0.2004 -0.4204 -0.0375
#> 
#> Columns 19 to 20  0.6776 -0.3055
#>   0.5942 -0.4065
#>   0.3361  0.0502
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.6854  0.6158  0.2599  0.4157 -0.5213  0.1942  0.5682  0.6990 -0.0695
#>   0.3248  0.3555  0.2364  0.4681  0.0206  0.4131  0.3864  0.0304  0.0601
#>   0.3837  0.2893  0.3941  0.6193 -0.1743  0.6747 -0.0794  0.7349  0.0661
#> 
#> Columns 10 to 18 -0.1248  0.0093 -0.3095  0.1268  0.2882 -0.1608  0.0559 -0.3806 -0.2431
#>   0.1041 -0.0540 -0.2657 -0.1043 -0.1668 -0.0254 -0.2020 -0.4176 -0.0031
#>  -0.3478 -0.2658 -0.3794  0.3883  0.1005 -0.2487 -0.0215 -0.6796  0.1072
#> 
#> Columns 19 to 20  0.7147  0.2637
#>   0.5439  0.0444
#>   0.6861  0.2690
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.3537  0.5238  0.4453  0.4405 -0.3047  0.4585  0.1731  0.5999  0.2031
#>   0.7216  0.5691  0.1289  0.4018 -0.0774 -0.1120  0.4812  0.2214 -0.3266
#>   0.4036  0.1039  0.4970  0.1151  0.0658  0.2532  0.4180 -0.0680 -0.1534
#> 
#> Columns 10 to 18 -0.6124 -0.2752  0.0525  0.1647  0.4035 -0.1581 -0.0362 -0.3190  0.2055
#>  -0.1584 -0.0231 -0.2413  0.2027  0.1127  0.0430  0.1294 -0.4220  0.1213
#>  -0.3172 -0.3751 -0.4767  0.3235  0.1889 -0.2880 -0.1723 -0.5751 -0.0795
#> 
#> Columns 19 to 20  0.8247  0.1197
#>   0.7547  0.3174
#>   0.5102 -0.3356
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.1055  0.1096  0.3701  0.2947  0.0575  0.6196 -0.3022  0.4294  0.0638
#>   0.2635  0.2058  0.3177 -0.0390 -0.1025  0.4044  0.0311  0.5352 -0.0115
#>   0.3382  0.0514  0.4530  0.3141 -0.2398  0.1567  0.1957  0.1010 -0.0640
#> 
#> Columns 10 to 18 -0.3804 -0.2116 -0.3530 -0.3659  0.1122 -0.2757 -0.3526 -0.5590  0.3342
#>  -0.1312 -0.1799 -0.2267 -0.0098 -0.0275  0.0285  0.1563 -0.6139 -0.2121
#>  -0.4800 -0.3502 -0.1726 -0.1388  0.0483 -0.3599 -0.0448 -0.1005 -0.1446
#> 
#> Columns 19 to 20  0.7925 -0.1909
#>   0.6036 -0.0115
#>   0.7342 -0.4311
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.4811 -0.7179  0.0980  0.1536  0.5751  0.6343 -0.4952  0.0682 -0.2995
#>  -0.3282 -0.1874 -0.2931  0.1595  0.5393  0.0219  0.3337 -0.2705 -0.6422
#>  -0.2632 -0.4486 -0.3421  0.3225 -0.3599  0.4799 -0.5787  0.2726 -0.5689
#> 
#> Columns 10 to 18 -0.3370 -0.5718  0.0451  0.4833 -0.0558 -0.3682  0.0413 -0.8035  0.8017
#>  -0.4908 -0.3515  0.4422  0.6361 -0.5180 -0.3570 -0.4741 -0.1396 -0.3848
#>   0.4833  0.0701  0.2339  0.3262 -0.1450  0.0176 -0.1722  0.3883  0.5110
#> 
#> Columns 19 to 20  0.4066 -0.4543
#>   0.0749 -0.3324
#>  -0.0835 -0.1594
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1055  0.1096  0.3701  0.2947  0.0575  0.6196 -0.3022  0.4294  0.0638
#>   0.2635  0.2058  0.3177 -0.0390 -0.1025  0.4044  0.0311  0.5352 -0.0115
#>   0.3382  0.0514  0.4530  0.3141 -0.2398  0.1567  0.1957  0.1010 -0.0640
#> 
#> Columns 10 to 18 -0.3804 -0.2116 -0.3530 -0.3659  0.1122 -0.2757 -0.3526 -0.5590  0.3342
#>  -0.1312 -0.1799 -0.2267 -0.0098 -0.0275  0.0285  0.1563 -0.6139 -0.2121
#>  -0.4800 -0.3502 -0.1726 -0.1388  0.0483 -0.3599 -0.0448 -0.1005 -0.1446
#> 
#> Columns 19 to 20  0.7925 -0.1909
#>   0.6036 -0.0115
#>   0.7342 -0.4311
#> [ CPUFloatType{2,3,20} ]
#>