RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2637  0.4769 -0.1880 -0.3161  0.2071  0.2447 -0.0992 -0.6145 -0.1850
#>   0.4432  0.0751 -0.2308 -0.1429 -0.3388  0.5710 -0.9610 -0.9772 -0.7058
#>   0.9148 -0.4575  0.5085  0.7993 -0.8019 -0.2457 -0.9439 -0.1951 -0.3417
#> 
#> Columns 10 to 18  0.3679 -0.5855 -0.4961 -0.4870 -0.4271  0.3752 -0.3193 -0.8350  0.8037
#>   0.6492  0.1396 -0.9292  0.9753  0.5135  0.5865  0.5975  0.9485  0.5268
#>   0.8514 -0.5117 -0.6721  0.6219 -0.3257  0.0904  0.4674 -0.0942  0.0967
#> 
#> Columns 19 to 20  0.3172 -0.3162
#>   0.1930 -0.6677
#>  -0.8948 -0.1229
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0807 -0.0749  0.4029 -0.1037 -0.0595  0.5732 -0.4772 -0.2078  0.2454
#>   0.3230  0.2950  0.2923 -0.1864 -0.5173  0.3354 -0.2861 -0.1460  0.4423
#>   0.6721 -0.4255  0.3732  0.2920 -0.1874 -0.0457 -0.0781  0.4600  0.5455
#> 
#> Columns 10 to 18  0.5738  0.5020  0.1810 -0.1060 -0.4189  0.3288  0.3715 -0.2230  0.0033
#>   0.3872 -0.1654 -0.5652 -0.3950 -0.4877  0.3189  0.7092 -0.0418 -0.0493
#>   0.1353 -0.2577 -0.1404 -0.0979 -0.5243  0.4489  0.2218 -0.7115  0.2924
#> 
#> Columns 19 to 20 -0.0160 -0.0630
#>  -0.0467  0.2756
#>  -0.6448 -0.6254
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5845  0.1329  0.4744  0.3723 -0.4245  0.1498 -0.6630 -0.1641  0.2018
#>   0.6606 -0.6317  0.4105  0.1522 -0.5623  0.4983 -0.5778  0.1133  0.1507
#>   0.6309 -0.5618  0.4491  0.1364  0.0075  0.3507 -0.4676  0.1441 -0.0601
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0134  0.7650  0.1439  0.4259  0.4153  0.6214  0.7226 -0.4236  0.2941
#>  -0.6863  0.8762  0.5323  0.3303 -0.6727 -0.5989  0.3618 -0.6909  0.4346
#>  -0.5951  0.1627 -0.4180  0.5354  0.0485 -0.1825  0.6306  0.4549  0.3979
#> 
#> Columns 10 to 18  0.6560  0.3911  0.1682 -0.1657  0.6458  0.5018  0.0587 -0.1725 -0.1522
#>   0.1087  0.4004 -0.6458 -0.7211 -0.2282 -0.3403 -0.6907  0.5181  0.8998
#>   0.2380  0.3833 -0.3462 -0.5668  0.4695  0.0495 -0.5703  0.1013  0.0684
#> 
#> Columns 19 to 20  0.3026 -0.1811
#>  -0.2732 -0.0296
#>   0.4510 -0.5177
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4715 -0.2208  0.4514  0.0689 -0.0362  0.1873 -0.0318  0.3327  0.6430
#>   0.2840 -0.5343  0.6533 -0.1517 -0.3876  0.0985 -0.6118  0.4881 -0.0187
#>   0.1703 -0.1130  0.4361 -0.1471 -0.1232  0.1931 -0.5729  0.1611  0.2231
#> 
#> Columns 10 to 18 -0.1548  0.0332  0.0542  0.4546 -0.2381  0.5078  0.0916 -0.2578  0.1670
#>  -0.2169 -0.0867  0.4298  0.1291 -0.1432  0.7374 -0.0253 -0.6034 -0.2054
#>   0.0406  0.0468  0.2289  0.1467 -0.0752  0.4663  0.1665 -0.4045  0.1724
#> 
#> Columns 19 to 20 -0.3480 -0.5057
#>  -0.4186 -0.6145
#>  -0.0279 -0.5708
#> [ CPUFloatType{2,3,20} ]
#>