RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0039  0.5477  0.1429 -0.0641 -0.6106  0.6007 -0.2357 -0.3873 -0.2844
#>   0.4412 -0.3140 -0.1049 -0.1538 -0.5370  0.9276  0.5927 -0.1612  0.5594
#>  -0.1189  0.6759  0.8197  0.6164  0.1025 -0.8686 -0.6821 -0.2801 -0.0580
#> 
#> Columns 10 to 18 -0.6116 -0.1496 -0.3106 -0.6787 -0.2787  0.5799 -0.0003 -0.1435  0.2070
#>  -0.5270  0.8306 -0.6020 -0.7831  0.0443 -0.2249  0.7407 -0.0843 -0.5417
#>  -0.8302 -0.0312  0.1526  0.4179 -0.0840 -0.4985 -0.9279  0.5034  0.8773
#> 
#> Columns 19 to 20  0.3655  0.8167
#>   0.8243 -0.0045
#>  -0.6862  0.2351
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2065 -0.4708  0.1815 -0.1216 -0.2686  0.3380 -0.2400 -0.2330 -0.1340
#>   0.0444  0.1011 -0.2721  0.8433  0.4459  0.3397  0.8063  0.7043 -0.2617
#>   0.1534 -0.7244 -0.5904 -0.1496 -0.0021 -0.0313  0.4016 -0.4670 -0.3529
#> 
#> Columns 10 to 18 -0.0405  0.2797 -0.4445 -0.2809  0.0259 -0.2130  0.0115 -0.0033  0.3094
#>   0.4315 -0.1642  0.0749 -0.6729 -0.3707 -0.1349  0.5115 -0.6538 -0.4410
#>  -0.4873  0.1298 -0.5018  0.5603 -0.0879  0.1096 -0.4852 -0.3304 -0.2207
#> 
#> Columns 19 to 20  0.4979 -0.2264
#>   0.3220 -0.5002
#>  -0.0830  0.0229
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1827  0.5177 -0.3944  0.5985  0.5983  0.4162  0.4843  0.2712 -0.4277
#>   0.2616  0.6388  0.1596 -0.1133  0.2286 -0.3054 -0.2143  0.2625  0.3499
#>   0.5536 -0.3101 -0.1050 -0.5760 -0.0892 -0.1124 -0.2508 -0.2112 -0.0574
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7352 -0.7738  0.1018 -0.3061 -0.3264  0.5255  0.2373  0.7276 -0.0656
#>   0.6468  0.0716 -0.3494  0.3424 -0.0383 -0.5537  0.2718  0.1965 -0.3363
#>   0.0757  0.4414 -0.2008  0.5313 -0.6390 -0.5607  0.2107 -0.4761  0.2476
#> 
#> Columns 10 to 18 -0.0865  0.0365  0.0924  0.2732 -0.0659 -0.2049 -0.5028  0.5528 -0.6238
#>   0.1423  0.1825 -0.7406  0.5041 -0.0896 -0.0498 -0.3110  0.0640 -0.2035
#>   0.4899 -0.2829 -0.5631  0.5330  0.5749  0.4930 -0.0748 -0.2331  0.6916
#> 
#> Columns 19 to 20  0.2012  0.4885
#>   0.3544 -0.1998
#>   0.2812  0.0640
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1965  0.0595 -0.2857  0.5515 -0.1028  0.5116  0.0937 -0.2257 -0.3296
#>   0.0902 -0.0454 -0.4177  0.4774  0.1664  0.2468 -0.1831  0.2617 -0.0557
#>   0.0243  0.1764 -0.3195  0.4498 -0.0680  0.1096  0.1243  0.3403 -0.2805
#> 
#> Columns 10 to 18 -0.0278 -0.3567 -0.0707  0.2779 -0.4523  0.1015 -0.3767  0.0881  0.0309
#>   0.1900 -0.4591 -0.2173  0.0963 -0.6295  0.4337  0.1291 -0.1971 -0.2753
#>   0.3808 -0.5508 -0.2336 -0.1972 -0.3484  0.4988  0.6300 -0.1025 -0.1588
#> 
#> Columns 19 to 20 -0.0052  0.0452
#>   0.2541  0.0864
#>   0.3598 -0.0020
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>