RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5804  0.3584 -0.8103  0.8278 -0.5261 -0.5261  0.3857 -0.9174 -0.6195
#>  -0.5433 -0.0864 -0.1534 -0.3547 -0.1379 -0.4889  0.4768 -0.3902 -0.5504
#>   0.6319  0.1976 -0.6662 -0.6248  0.3967  0.2512 -0.3049 -0.3077  0.3196
#> 
#> Columns 10 to 18 -0.0243  0.4637  0.4064 -0.2444 -0.2915  0.3553 -0.9199 -0.4118  0.1102
#>   0.1513  0.5802  0.6446  0.0770 -0.2088 -0.7415 -0.0055  0.3487 -0.2016
#>  -0.0620  0.6022  0.2617  0.1674  0.5754 -0.4503  0.1062 -0.2600 -0.2486
#> 
#> Columns 19 to 20 -0.4137  0.4736
#>  -0.8385 -0.6101
#>   0.2520  0.2983
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1618  0.0927  0.4738  0.1726  0.1962 -0.4565 -0.4422  0.1430 -0.4333
#>   0.5175 -0.2608 -0.0271 -0.3246 -0.1705  0.3484  0.1372  0.4279 -0.2210
#>  -0.0091 -0.0467 -0.4164  0.2221  0.2976  0.2803  0.1958 -0.3322 -0.7979
#> 
#> Columns 10 to 18  0.0045  0.5655  0.0552  0.4740 -0.5064 -0.0551 -0.2805 -0.1812  0.6923
#>  -0.1254  0.2893 -0.3301  0.0206 -0.6549  0.0230 -0.4854  0.1750  0.2206
#>  -0.1169  0.1597  0.0166 -0.0488  0.2109  0.5101 -0.5959  0.2240  0.2783
#> 
#> Columns 19 to 20 -0.1762 -0.0523
#>  -0.1946 -0.5285
#>  -0.2807 -0.7736
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1909 -0.4878  0.0063  0.0284  0.0209  0.3138  0.0092 -0.1831  0.0232
#>  -0.5219  0.2915  0.3243  0.6016 -0.1531 -0.2941 -0.1324  0.0970 -0.2094
#>   0.0729 -0.0352  0.2471  0.3101 -0.2709  0.1405 -0.4948  0.2572 -0.0929
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0209 -0.4747  0.5467  0.3920 -0.1172 -0.5229  0.3578 -0.7114 -0.2937
#>  -0.8003 -0.3472  0.3022 -0.4246 -0.2637  0.0780 -0.0028 -0.7777  0.5391
#>   0.8670  0.7291 -0.2709  0.6770  0.0711 -0.5232 -0.0812 -0.6359  0.6428
#> 
#> Columns 10 to 18 -0.3656  0.1397  0.1792 -0.0764  0.5382 -0.5254  0.7756 -0.4427  0.7114
#>  -0.6022  0.4279 -0.0855  0.6066 -0.3248 -0.0223  0.4846 -0.1512  0.3631
#>  -0.6652  0.5104  0.6817 -0.6264  0.3854 -0.6313  0.5315 -0.8061 -0.4137
#> 
#> Columns 19 to 20 -0.7381  0.5038
#>  -0.5777 -0.0472
#>  -0.6883  0.1382
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1433  0.2199  0.2288  0.4580 -0.2422 -0.1434 -0.3635 -0.3478 -0.3321
#>  -0.2272 -0.0169  0.0216  0.3757 -0.1976 -0.1856  0.0368 -0.0762 -0.2004
#>  -0.0067 -0.0155 -0.1525 -0.3309 -0.1417  0.3070 -0.0467  0.3593 -0.0023
#> 
#> Columns 10 to 18  0.0267  0.1197  0.0044 -0.0966 -0.2871 -0.1795 -0.5820  0.3531  0.5290
#>  -0.2991  0.5461  0.1739  0.0706 -0.2119 -0.2141 -0.5018 -0.1568  0.4076
#>  -0.1784  0.7053 -0.1596 -0.0051 -0.2371 -0.0332  0.3879  0.4322  0.3243
#> 
#> Columns 19 to 20  0.1533 -0.3379
#>  -0.3103 -0.5454
#>  -0.0925 -0.0548
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>