RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4893 -0.5359 -0.0320  0.8564  0.3143  0.4615 -0.1793  0.5171 -0.3303
#>   0.3498 -0.6527 -0.3970  0.2764  0.5832  0.7689  0.2603  0.2732 -0.4107
#>  -0.7191  0.3119 -0.2592 -0.8349  0.1389  0.1983  0.2625 -0.3058 -0.4589
#> 
#> Columns 10 to 18 -0.4410  0.7786 -0.3830  0.4595  0.6183  0.3650 -0.2805  0.6443 -0.1254
#>  -0.6176  0.1930 -0.4548 -0.2538 -0.0121  0.4840  0.4179 -0.2716 -0.6813
#>  -0.7103  0.6698 -0.2196 -0.5443  0.5020 -0.5175  0.1793 -0.8579 -0.2832
#> 
#> Columns 19 to 20 -0.7283 -0.1427
#>   0.2094 -0.3939
#>  -0.1572 -0.6349
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3757  0.0891 -0.6864  0.3380  0.8685 -0.1697 -0.1614  0.4526 -0.2175
#>  -0.1934  0.2901 -0.1768 -0.1153  0.3144 -0.3449  0.1018  0.4583 -0.1932
#>   0.3778  0.4473 -0.6042 -0.0630  0.5540 -0.0350  0.2483  0.0087 -0.3715
#> 
#> Columns 10 to 18  0.1646 -0.1009  0.5457 -0.1617  0.5780 -0.0846  0.0619  0.3084  0.6082
#>  -0.5816  0.5230 -0.4328 -0.2459  0.4778  0.0182  0.1209  0.3413  0.4606
#>  -0.7588  0.0971  0.4782  0.0658 -0.1272 -0.2059 -0.1252 -0.0858  0.0525
#> 
#> Columns 19 to 20 -0.3043  0.3272
#>  -0.1738 -0.0024
#>   0.0826  0.4984
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0835 -0.1242 -0.1694  0.3785  0.4170  0.1485 -0.5281  0.1856  0.1038
#>   0.0274  0.5230 -0.6051  0.2048  0.8159 -0.0723 -0.1515  0.2752 -0.1904
#>   0.3274 -0.2028 -0.6917  0.4267  0.3736  0.0728 -0.4883  0.2301 -0.1851
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7280 -0.4662  0.3045 -0.6991  0.0021 -0.3212 -0.0428  0.8022 -0.3606
#>   0.7928  0.1978 -0.4499 -0.6096 -0.8820  0.0712 -0.5038  0.8133 -0.5525
#>  -0.0886  0.6213 -0.0903  0.1403  0.1276 -0.6544  0.1016  0.0047  0.0948
#> 
#> Columns 10 to 18  0.3555  0.7347  0.4678  0.3534 -0.1461  0.4836  0.5409 -0.4846 -0.8273
#>   0.5893  0.5701  0.5243  0.6558  0.0963  0.5138  0.6117 -0.3865 -0.6832
#>  -0.3077  0.0723  0.3993  0.7193  0.1465 -0.1073  0.0520  0.0208 -0.0481
#> 
#> Columns 19 to 20 -0.3718  0.8337
#>   0.2928  0.9177
#>   0.6285  0.5206
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3341 -0.1495  0.0247  0.2134  0.4671 -0.1561 -0.1576 -0.2701 -0.3562
#>  -0.1476  0.2792 -0.3159  0.3663  0.6435 -0.5657 -0.2347  0.1601 -0.3295
#>  -0.1968  0.1553 -0.1089 -0.0865  0.1013 -0.2072 -0.0832  0.3615 -0.4205
#> 
#> Columns 10 to 18 -0.7854  0.0782 -0.2375  0.1675 -0.0431 -0.1840 -0.2008  0.2532  0.0804
#>  -0.7010  0.2477  0.2367  0.3089  0.2298 -0.1819 -0.2643  0.6033  0.1311
#>  -0.4275  0.2244  0.2684 -0.0832  0.3501  0.3810  0.2262  0.2812  0.4939
#> 
#> Columns 19 to 20  0.2855  0.3009
#>   0.2894  0.1285
#>   0.0798  0.2501
#> [ CPUFloatType{2,3,20} ]
#>