RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7204  0.8010  0.0852  0.4051  0.8662 -0.0600  0.1933  0.5523  0.7757
#>  -0.3983 -0.0762 -0.6636  0.0649 -0.2439  0.3113 -0.3695  0.5917  0.6844
#>   0.5036  0.0436 -0.3274 -0.0512  0.9097  0.2355  0.2896  0.0359  0.1615
#> 
#> Columns 10 to 18  0.6783  0.2410 -0.6612  0.3673  0.0559 -0.6603  0.0200 -0.6973 -0.0581
#>  -0.2302 -0.1901  0.2528  0.4080 -0.5817  0.6220  0.3088  0.6923  0.0225
#>   0.6839  0.2628 -0.2174  0.3426 -0.8628 -0.3861  0.5080  0.0406 -0.0411
#> 
#> Columns 19 to 20  0.5002  0.1293
#>   0.5803  0.1123
#>  -0.7370  0.2447
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2105  0.4546  0.2348 -0.3237  0.0673  0.1333 -0.1755  0.2644 -0.4330
#>   0.0972 -0.4628  0.3643 -0.6638 -0.0325 -0.4539  0.3944  0.5462  0.3472
#>   0.0546  0.1708  0.6673 -0.4571 -0.1804  0.3708  0.1974  0.6736 -0.2363
#> 
#> Columns 10 to 18 -0.1870  0.2385  0.0434  0.1382 -0.0917  0.0673 -0.4800  0.4988  0.6180
#>  -0.0128  0.3655  0.1678  0.2836 -0.1690  0.0766  0.6275  0.5056 -0.1079
#>   0.0148 -0.4127  0.3637  0.0024 -0.4033 -0.2609  0.5150  0.5684  0.0286
#> 
#> Columns 19 to 20 -0.2786 -0.5948
#>  -0.2819 -0.0260
#>   0.0997  0.1585
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0175  0.2552 -0.1879 -0.3985 -0.4290  0.1330 -0.0677  0.1084  0.4420
#>  -0.1370 -0.2764  0.2743 -0.6698  0.3328  0.1035  0.6550  0.6464  0.2383
#>   0.1194 -0.0110  0.3027 -0.6372  0.2449  0.2152 -0.0678  0.3391 -0.0058
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0744  0.1439  0.4529  0.4960  0.3836  0.6220 -0.0229  0.0382 -0.3584
#>  -0.1835  0.3608  0.5774 -0.2005 -0.7740  0.0942  0.5867 -0.3646  0.3440
#>  -0.1004 -0.4553 -0.7341  0.0962  0.0726 -0.1463  0.4169  0.7298  0.5383
#> 
#> Columns 10 to 18  0.6648  0.5176 -0.0087 -0.1661 -0.3157  0.2492  0.4310 -0.3061  0.5885
#>   0.0455 -0.2686  0.0507 -0.4120 -0.6909  0.1927  0.3917 -0.3171  0.4085
#>   0.4740  0.2643 -0.7732  0.0817  0.2000  0.7951 -0.0070  0.3304 -0.6799
#> 
#> Columns 19 to 20 -0.3429 -0.0675
#>  -0.0193  0.3441
#>   0.4855 -0.5135
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4443  0.1542  0.2674 -0.6907  0.4651 -0.0628  0.3187  0.2006 -0.1586
#>   0.5231  0.6378  0.2588 -0.1699  0.4660  0.5783  0.2729  0.4857 -0.0869
#>   0.3168 -0.2100  0.3591 -0.5516 -0.3173  0.2053  0.3210  0.3299  0.4991
#> 
#> Columns 10 to 18 -0.1290 -0.2106  0.1432 -0.1203 -0.1307 -0.2073  0.3484  0.6623 -0.0190
#>  -0.2810 -0.0152 -0.0878  0.5767 -0.6310  0.0283  0.0888  0.4012  0.2129
#>  -0.2252 -0.2841  0.1641  0.7643 -0.1349  0.4327  0.4270 -0.0095  0.3655
#> 
#> Columns 19 to 20 -0.3434 -0.3478
#>   0.1587 -0.0509
#>  -0.2965 -0.2491
#> [ CPUFloatType{2,3,20} ]
#>