RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3762 -0.0498  0.2931  0.4407 -0.3580  0.5962  0.9029  0.3094  0.0283
#>   0.6458  0.6209 -0.1792  0.3720  0.6428 -0.5651  0.6146  0.5848  0.3791
#>  -0.7399  0.7700 -0.4092  0.5679  0.0460  0.6964  0.5561  0.3577 -0.5670
#> 
#> Columns 10 to 18 -0.4902 -0.8050  0.3296 -0.1186 -0.4642  0.5700  0.4451  0.9252 -0.9083
#>  -0.5221 -0.2399 -0.3480 -0.1102 -0.5717  0.7630 -0.4341  0.7801 -0.7407
#>   0.4780  0.5990  0.1520 -0.2659  0.2131  0.1960  0.4185 -0.5447  0.8726
#> 
#> Columns 19 to 20  0.6677  0.6729
#>   0.6694 -0.0581
#>   0.4613 -0.8599
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0124  0.1432  0.3892 -0.0960 -0.4957  0.2701  0.3803  0.0557 -0.1103
#>   0.0657  0.0512  0.2482  0.5363 -0.0092 -0.4451 -0.0333 -0.0322  0.3405
#>   0.3363  0.1828 -0.2277  0.3527  0.0684  0.2894  0.3068 -0.3521  0.5823
#> 
#> Columns 10 to 18 -0.0645  0.2452  0.6435 -0.2297 -0.1589  0.2237 -0.1352  0.3991  0.3823
#>  -0.0122  0.1171  0.5416 -0.4471 -0.1682  0.6414 -0.3613 -0.0146  0.0034
#>  -0.6542 -0.0766 -0.3002  0.5012 -0.1372  0.1250 -0.1117 -0.4786 -0.3567
#> 
#> Columns 19 to 20  0.0752  0.0423
#>  -0.4489 -0.1154
#>   0.0485 -0.7820
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2013  0.5774  0.4083  0.0506  0.0527  0.3584  0.3428 -0.2323 -0.3890
#>  -0.0106  0.3927  0.5273 -0.2459 -0.2219  0.1232  0.2567 -0.1832  0.4173
#>   0.6110  0.2243  0.3759  0.3913  0.5091  0.1933 -0.3137 -0.1702  0.0340
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3066  0.2529 -0.6372 -0.3074 -0.1674 -0.0284 -0.2668  0.0036 -0.4458
#>   0.0847 -0.0320 -0.5672  0.1819 -0.4201  0.3028 -0.1372  0.4160 -0.4026
#>   0.4453  0.4821 -0.7387 -0.2470  0.2262 -0.5341 -0.5004 -0.7837  0.2142
#> 
#> Columns 10 to 18 -0.4629  0.1865  0.3940 -0.1388 -0.2684  0.2491 -0.5534  0.1627  0.5763
#>  -0.4806  0.5748  0.1142 -0.0165  0.0287 -0.1719 -0.3876  0.0453  0.1649
#>  -0.6599  0.1287  0.4004  0.3965 -0.2518 -0.3625  0.0349 -0.1488 -0.0560
#> 
#> Columns 19 to 20 -0.5034  0.4683
#>  -0.3908  0.6101
#>  -0.1599  0.2212
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5863  0.4502  0.3403  0.4298 -0.0283  0.2576 -0.1963 -0.1526  0.1800
#>   0.3181  0.5425  0.2992  0.2113  0.0135  0.2147 -0.1761  0.2098 -0.1125
#>   0.2501  0.2050  0.1429  0.3658  0.3339  0.4044 -0.3869 -0.3779  0.4500
#> 
#> Columns 10 to 18 -0.3802  0.2696 -0.0684 -0.3775 -0.1351  0.3636 -0.2455  0.1268  0.1679
#>  -0.2521  0.2495 -0.0425 -0.4018 -0.3545  0.2357 -0.2502  0.2182  0.1708
#>  -0.3042  0.3638  0.2086  0.1453 -0.1243  0.5234 -0.4833  0.0725 -0.1572
#> 
#> Columns 19 to 20  0.2942 -0.1516
#>   0.4161 -0.0556
#>  -0.1780 -0.0355
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>