RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6941  0.3050  0.2629  0.8123 -0.5735 -0.0215 -0.4435 -0.0733  0.1482
#>   0.1653 -0.5385 -0.5675  0.1460 -0.4904  0.1470  0.4018  0.2520 -0.1343
#>  -0.0668  0.3298  0.5395 -0.1108 -0.0575  0.5604 -0.1399  0.5457  0.4014
#> 
#> Columns 10 to 18  0.2857  0.7130 -0.3147  0.1116  0.0306 -0.8288  0.6984  0.4902  0.8663
#>   0.4635 -0.3858 -0.3649 -0.1907  0.4397 -0.5303  0.3004  0.4190  0.3705
#>   0.0367  0.4021  0.0900  0.3193 -0.0757  0.1649  0.1506  0.2770 -0.8533
#> 
#> Columns 19 to 20 -0.7995  0.0589
#>  -0.3452 -0.6080
#>  -0.2347 -0.2998
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2782 -0.1019  0.5690 -0.1185 -0.3255  0.0681 -0.4487  0.0824  0.3216
#>   0.0641 -0.0949  0.1001 -0.2960 -0.3967  0.2753 -0.6685 -0.1652  0.2911
#>   0.2761 -0.1036  0.3373  0.0362 -0.4439  0.2330  0.0444  0.0708 -0.0766
#> 
#> Columns 10 to 18  0.5776 -0.5942 -0.1758  0.6639 -0.2389 -0.5198  0.3414  0.3047 -0.5873
#>   0.1863 -0.2368  0.1123  0.6227 -0.0684 -0.3314  0.1831  0.4322  0.1818
#>   0.7153 -0.5697 -0.2102  0.4996  0.0726 -0.0917  0.3432  0.3488  0.1606
#> 
#> Columns 19 to 20 -0.7363  0.8613
#>   0.5386  0.4210
#>  -0.2039  0.6454
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4951 -0.1392  0.5692  0.2143 -0.5981  0.0822 -0.0616 -0.2765  0.3981
#>   0.3028 -0.2552  0.2147  0.0806 -0.4727  0.2587 -0.7178  0.5890 -0.0432
#>   0.1389 -0.0605 -0.1419 -0.1968 -0.2948 -0.2504 -0.0341  0.3954  0.3121
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4911  0.3755  0.4751 -0.6065  0.5340  0.3392 -0.0575 -0.6347 -0.1480
#>   0.5731 -0.5228 -0.1515 -0.4468 -0.1293  0.1355 -0.4412  0.2075 -0.0126
#>   0.8260  0.2436  0.7347  0.3410  0.5267  0.1694  0.0403  0.3366 -0.5699
#> 
#> Columns 10 to 18 -0.4047  0.0636 -0.1738  0.4979  0.2563 -0.2617 -0.2884 -0.5654  0.4045
#>  -0.2726  0.1614  0.0045 -0.2271 -0.0160  0.0834 -0.4575 -0.5347  0.0685
#>  -0.1330  0.8864 -0.0070 -0.7806 -0.3543  0.3368 -0.2984 -0.5998 -0.4124
#> 
#> Columns 19 to 20  0.3251 -0.7676
#>   0.4646 -0.0382
#>   0.5053 -0.7112
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1174  0.2234 -0.0523 -0.0605 -0.3522  0.2682 -0.2854  0.0444  0.2182
#>   0.2287  0.1125  0.2196  0.1073 -0.1507  0.1420 -0.4910 -0.2752  0.1795
#>   0.4029  0.0151  0.4247 -0.4080 -0.5988  0.3850 -0.6597 -0.1840  0.1628
#> 
#> Columns 10 to 18  0.7806 -0.0429 -0.4840  0.6682 -0.7598 -0.3945  0.6685  0.1158  0.2570
#>   0.4776 -0.0776 -0.3394  0.5918 -0.2772 -0.3582  0.0155  0.6222  0.3519
#>   0.5888 -0.3352 -0.0171  0.4368 -0.6569 -0.3781  0.3847  0.1560 -0.0708
#> 
#> Columns 19 to 20 -0.5485  0.4045
#>  -0.0722  0.3309
#>  -0.4542  0.6721
#> [ CPUFloatType{2,3,20} ]
#>