RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1364 -0.6365 -0.0160  0.9638 -0.7737 -0.2509 -0.6532 -0.5614  0.4514
#>   0.4927 -0.3643 -0.5249  0.2017  0.3069  0.5998  0.7679  0.2964  0.0982
#>  -0.6095  0.2111  0.2673  0.5243 -0.5791  0.5881  0.2470 -0.0806  0.2366
#> 
#> Columns 10 to 18  0.6218  0.1493  0.7587  0.3080 -0.3386 -0.6210  0.3623  0.1924 -0.4969
#>  -0.8693  0.8508 -0.7011 -0.4280 -0.3342  0.1343 -0.1945 -0.3650 -0.4178
#>  -0.7838  0.4238 -0.7747 -0.8476  0.1490 -0.0695 -0.3372 -0.5311 -0.2250
#> 
#> Columns 19 to 20  0.1048 -0.7276
#>  -0.8982 -0.4776
#>  -0.2337  0.5778
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2788 -0.1759  0.1854  0.1319  0.2074  0.3401  0.0691 -0.4653  0.1860
#>  -0.0572  0.1809  0.2645  0.7114 -0.1092  0.0580  0.1569  0.3942 -0.3044
#>   0.5116  0.1972 -0.3033 -0.3125  0.3592 -0.3356  0.2860  0.4385 -0.3397
#> 
#> Columns 10 to 18  0.2491 -0.6468 -0.2423 -0.2515  0.2635 -0.0747  0.1124  0.0834  0.3543
#>   0.5622  0.5982  0.5045  0.4186  0.6172  0.2331  0.7911 -0.4940 -0.7462
#>   0.1019  0.4698  0.2174  0.7002  0.2105 -0.4304  0.1212 -0.7463 -0.4263
#> 
#> Columns 19 to 20 -0.0036  0.0350
#>  -0.3673 -0.2588
#>   0.1235 -0.5994
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2406  0.1716 -0.2910  0.1525  0.0391 -0.1983  0.4527 -0.1546  0.1318
#>  -0.2844 -0.0827  0.4214  0.3852  0.0852  0.2195 -0.2573  0.1230 -0.4211
#>  -0.3943 -0.2101  0.2156  0.5000  0.1351  0.1795 -0.1387 -0.0614 -0.1637
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5356  0.5133  0.0062  0.7102 -0.1049  0.1931 -0.1430 -0.4652 -0.0035
#>   0.1215 -0.4529 -0.1149 -0.1915  0.2933 -0.7744  0.1713  0.2055 -0.8370
#>   0.4246 -0.0566  0.2314  0.2184 -0.4111 -0.5511  0.1661  0.0257 -0.4286
#> 
#> Columns 10 to 18 -0.2716  0.4798  0.4141  0.0044  0.0606 -0.6382  0.2693 -0.6086 -0.5589
#>   0.1921 -0.0159 -0.0122  0.6196  0.2068  0.2132 -0.7669 -0.0346  0.7527
#>   0.2550 -0.0936 -0.1568  0.1018 -0.1633 -0.3870 -0.1093 -0.1775  0.6577
#> 
#> Columns 19 to 20  0.3020 -0.0915
#>  -0.3016  0.6092
#>   0.3338  0.8669
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4099  0.0573  0.0694  0.2523  0.1858  0.2455 -0.1111  0.1304 -0.5367
#>   0.0026  0.1525 -0.1350 -0.1306  0.2899  0.0323  0.0981 -0.0539  0.4448
#>   0.2160 -0.2344 -0.5166  0.1286  0.2073  0.0110  0.1358 -0.0675  0.0685
#> 
#> Columns 10 to 18 -0.2154  0.5333 -0.2262  0.1683  0.3503 -0.2268  0.1276 -0.5524  0.3506
#>   0.1223 -0.3325 -0.0239  0.1111  0.2865 -0.3596  0.1342 -0.1990  0.1563
#>  -0.1995  0.1883 -0.3816  0.3560  0.1720 -0.0271  0.2831 -0.6360  0.0456
#> 
#> Columns 19 to 20  0.4205 -0.4631
#>  -0.0956 -0.3495
#>  -0.3353 -0.0810
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>