RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2031 -0.0926  0.3078 -0.7228 -0.2030  0.1105 -0.4294 -0.3705  0.8229
#>  -0.2427  0.0975 -0.7399  0.2560 -0.5968 -0.6697 -0.1329 -0.1889 -0.3816
#>  -0.2393  0.4230 -0.8563 -0.6164 -0.1951  0.4147 -0.5119 -0.9378 -0.4069
#> 
#> Columns 10 to 18 -0.7219  0.6538 -0.4456  0.6718  0.7813  0.4203 -0.1207 -0.7912 -0.1597
#>   0.4278 -0.2460  0.6189 -0.6360 -0.1227 -0.0674  0.0954  0.5449 -0.2552
#>  -0.6103  0.6308  0.0670 -0.0329 -0.3564 -0.6823 -0.3614 -0.5316 -0.5317
#> 
#> Columns 19 to 20 -0.5925 -0.1632
#>   0.7832 -0.4283
#>   0.2199 -0.6287
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3916  0.1275 -0.2096 -0.2475 -0.4383 -0.2763 -0.0962 -0.7544 -0.0198
#>  -0.3987  0.2592 -0.3264  0.0937  0.1873 -0.0135 -0.2660 -0.6719 -0.2275
#>  -0.5157 -0.0845  0.1590  0.2091 -0.1164 -0.4512  0.1718 -0.3785 -0.3933
#> 
#> Columns 10 to 18 -0.1699 -0.6481  0.0925 -0.6853 -0.2090 -0.6360 -0.0616  0.4147 -0.3154
#>  -0.0600  0.3160 -0.2002 -0.0729  0.6007  0.2790 -0.6158  0.3083 -0.0484
#>  -0.4633  0.4289 -0.2008 -0.5681  0.6822 -0.1377 -0.2983  0.2441 -0.2401
#> 
#> Columns 19 to 20 -0.6273 -0.3290
#>   0.1798 -0.3575
#>  -0.3422  0.2222
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0471  0.7082  0.0075  0.4067  0.1691  0.1303  0.0954 -0.3880 -0.2446
#>  -0.1819 -0.2566 -0.1978 -0.2956 -0.1643 -0.1889 -0.0737 -0.5756 -0.3813
#>   0.3961 -0.2693  0.1181 -0.4124  0.1331 -0.2006  0.2164 -0.7346 -0.0599
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2545  0.2012 -0.3814 -0.1212  0.5074  0.1647 -0.4870  0.4001 -0.2777
#>   0.3767  0.0543 -0.4319 -0.0542 -0.6636 -0.2402 -0.2456  0.2580  0.1476
#>  -0.2470  0.4952 -0.2737  0.3486 -0.3561  0.5813 -0.0034  0.1261 -0.5999
#> 
#> Columns 10 to 18 -0.3945 -0.3427  0.5776 -0.0282 -0.4283  0.0424 -0.0892  0.2344  0.0306
#>  -0.1741  0.4039 -0.1020 -0.2529 -0.2469  0.1041  0.5978 -0.0834 -0.0810
#>  -0.0778  0.2960  0.6550 -0.6307 -0.6916  0.5768  0.1668  0.0176 -0.6425
#> 
#> Columns 19 to 20  0.0482 -0.5183
#>   0.1125 -0.6416
#>  -0.1436 -0.4155
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0194  0.3680 -0.2890 -0.2628 -0.1248 -0.0038 -0.0533 -0.3690  0.1867
#>  -0.0753  0.0352 -0.2375 -0.2425  0.2786  0.0584 -0.1588 -0.0931 -0.2175
#>  -0.1965  0.4301 -0.3709 -0.0001 -0.2687  0.1165  0.1325 -0.1822 -0.1604
#> 
#> Columns 10 to 18 -0.1235  0.0407 -0.0758 -0.4085  0.2125 -0.3667 -0.3091 -0.2627  0.0222
#>  -0.4618  0.5404  0.0918 -0.2977  0.3908  0.3265 -0.5179 -0.3697  0.3362
#>   0.0264  0.1955 -0.4032 -0.3323  0.3689 -0.3705 -0.3066 -0.5459  0.3484
#> 
#> Columns 19 to 20 -0.2460 -0.3354
#>   0.1561 -0.2663
#>  -0.3639 -0.1315
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>