RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0752  0.0780 -0.6489 -0.2852  0.2348  0.2037  0.2091 -0.4044 -0.3485
#>   0.8456  0.2214  0.1860  0.3173 -0.4928  0.0786 -0.2514  0.0282 -0.0166
#>  -0.7911  0.7263  0.3046 -0.0190 -0.7971  0.4043 -0.3926  0.5363 -0.1507
#> 
#> Columns 10 to 18 -0.0514 -0.4872 -0.4899  0.8021 -0.2896  0.6097  0.1580 -0.0051 -0.4361
#>   0.1222 -0.7343  0.2688  0.2968 -0.1649 -0.1632  0.3247 -0.2436 -0.3153
#>   0.1188 -0.7536 -0.2071  0.5977 -0.5947  0.9220 -0.7587 -0.1046 -0.5665
#> 
#> Columns 19 to 20 -0.3610 -0.4967
#>  -0.5347 -0.1991
#>  -0.0742 -0.2299
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2641  0.2814 -0.6586  0.0553 -0.3986  0.0996 -0.3435  0.3286 -0.4218
#>  -0.3376  0.5635 -0.1564  0.1472 -0.6538 -0.3041 -0.2629 -0.0892  0.5834
#>   0.4390  0.3138  0.4367 -0.0955 -0.0759 -0.2590 -0.0482  0.0103 -0.3886
#> 
#> Columns 10 to 18  0.4909  0.0578 -0.6368  0.0807 -0.2715 -0.0479 -0.3369  0.1348 -0.2074
#>   0.0305  0.1751 -0.0965  0.2521 -0.4905  0.3259 -0.6000 -0.0750  0.3353
#>  -0.2424  0.0272 -0.5603 -0.0892 -0.3368  0.1596  0.2759 -0.0484  0.1027
#> 
#> Columns 19 to 20 -0.2208 -0.1640
#>   0.3846 -0.4528
#>  -0.4051  0.2332
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1221  0.3041 -0.2937  0.0604 -0.2898 -0.4446 -0.0786  0.1583 -0.1179
#>  -0.2048  0.1908  0.1925  0.2072 -0.0264 -0.0865  0.0757 -0.2550  0.1037
#>  -0.2637  0.3230  0.0650 -0.3947  0.1495  0.0589 -0.2906  0.1950 -0.0900
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2279 -0.6432 -0.4972 -0.0291 -0.1592  0.0812 -0.2645 -0.5092 -0.0952
#>  -0.8563  0.5751  0.5921  0.6337  0.7662 -0.4044 -0.7343  0.4717  0.1260
#>  -0.1731  0.3862  0.3096  0.5986 -0.1467  0.4532 -0.2603  0.6128  0.4081
#> 
#> Columns 10 to 18 -0.1265  0.0102 -0.2973  0.3003  0.1708  0.3081 -0.5957  0.3477  0.3876
#>   0.1590 -0.1198  0.5489  0.1295  0.8045  0.4644 -0.5968  0.6829  0.4887
#>   0.3745  0.0830 -0.0310  0.0203 -0.1746 -0.2196 -0.5729  0.5550  0.6078
#> 
#> Columns 19 to 20  0.8235  0.0377
#>  -0.4533  0.1326
#>   0.4427  0.4523
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1289 -0.1075 -0.5863 -0.0914 -0.3443 -0.0364 -0.6298  0.3019  0.2945
#>  -0.0462  0.5011  0.1016  0.5178 -0.1874  0.0564  0.4229 -0.4633 -0.0216
#>  -0.6178  0.3044 -0.1643 -0.1484  0.1702 -0.2663  0.0563 -0.3998 -0.2791
#> 
#> Columns 10 to 18  0.0145 -0.2616 -0.4254  0.1217 -0.1691 -0.4266 -0.6401 -0.0447  0.1793
#>  -0.1501 -0.2007 -0.0208  0.3269 -0.2489  0.3874 -0.2068 -0.1985  0.4033
#>  -0.6488  0.0203  0.0498  0.0004 -0.2881  0.3977 -0.4904 -0.1783  0.0350
#> 
#> Columns 19 to 20  0.5120 -0.2003
#>   0.4213 -0.4882
#>   0.1786 -0.5776
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>