RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7643  0.4751 -0.3851  0.4440  0.7057  0.5772 -0.1788  0.0811 -0.4599
#>   0.7263 -0.1246  0.7757 -0.4246 -0.0671  0.0830  0.2346 -0.0961  0.1642
#>  -0.5465 -0.3270  0.1765  0.9038  0.1369  0.5544  0.4490  0.4902 -0.6345
#> 
#> Columns 10 to 18 -0.5543 -0.5864 -0.4810 -0.6776  0.0322 -0.1834  0.6051 -0.2329  0.5151
#>   0.2451  0.1542  0.0451 -0.2525  0.8905 -0.0236  0.1965  0.4556  0.5111
#>  -0.5763 -0.6692  0.1179  0.8161 -0.1782  0.2358  0.3680  0.4564  0.0868
#> 
#> Columns 19 to 20  0.3330  0.8843
#>  -0.2970 -0.6919
#>  -0.6878 -0.3558
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7802  0.2004  0.0220 -0.5137 -0.5729  0.0357  0.3821  0.1138  0.4851
#>  -0.5712  0.0164  0.2424 -0.3465  0.2516  0.1265  0.4078  0.1095 -0.3999
#>  -0.6437  0.2264 -0.1130 -0.4957 -0.8626 -0.5612  0.5339 -0.1458  0.3406
#> 
#> Columns 10 to 18 -0.0335 -0.1992  0.2153  0.1022  0.0442  0.5578  0.4253 -0.1442  0.3730
#>  -0.2205 -0.1834  0.3461 -0.1884  0.4432 -0.4162  0.2888  0.1600  0.4214
#>   0.0976 -0.1868 -0.3603  0.3098  0.0591  0.2063  0.0246  0.1961  0.4541
#> 
#> Columns 19 to 20  0.0132  0.3304
#>   0.0582  0.2725
#>  -0.3027 -0.0549
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3994  0.0343  0.2754 -0.1018  0.0168  0.1112 -0.0074 -0.0999 -0.1212
#>  -0.5943  0.1341  0.4957 -0.5094 -0.3669 -0.0762  0.4452  0.0352  0.4063
#>  -0.3679 -0.1663  0.2108 -0.3263 -0.3316  0.2431 -0.0531 -0.0050 -0.4928
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1589 -0.2573  0.5222 -0.1744 -0.7037  0.1147  0.6175  0.0569 -0.5612
#>   0.1780 -0.0785 -0.4189 -0.1113 -0.2250  0.1839  0.6337 -0.6501 -0.0816
#>   0.1351 -0.2114  0.6691  0.4766  0.3112  0.1294 -0.3085 -0.1096 -0.1372
#> 
#> Columns 10 to 18 -0.1502 -0.4481 -0.0159  0.2560 -0.5102  0.7256 -0.6431  0.1257 -0.5565
#>  -0.2130  0.7544  0.5091 -0.4801 -0.0469  0.3772 -0.3027 -0.2653  0.3002
#>   0.1771 -0.0734  0.6278  0.1449  0.1030  0.2079 -0.1248 -0.6378  0.2880
#> 
#> Columns 19 to 20 -0.7072  0.0560
#>  -0.2668 -0.0261
#>  -0.2897  0.3944
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4793  0.1549 -0.2281 -0.3639 -0.1333 -0.1747  0.5461  0.5028 -0.0156
#>  -0.7055 -0.0951  0.0214 -0.2881  0.1250  0.0139  0.3052  0.0372 -0.1252
#>  -0.4834  0.3334 -0.4384 -0.2412 -0.1297 -0.1174  0.7289  0.4035 -0.0315
#> 
#> Columns 10 to 18 -0.1966  0.5901  0.6761 -0.5218  0.0470 -0.6349  0.4548  0.0604  0.5200
#>  -0.0197 -0.1577  0.2184 -0.1092  0.1303 -0.0350 -0.2242  0.3443 -0.1194
#>   0.4007 -0.2502  0.6214 -0.1282  0.2997 -0.3858  0.3126  0.1875  0.3401
#> 
#> Columns 19 to 20 -0.1603  0.3401
#>   0.1501  0.1243
#>  -0.2710 -0.0272
#> [ CPUFloatType{2,3,20} ]
#>