RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2114 -0.6742 -0.5311  0.6799 -0.0536 -0.8183 -0.6923 -0.5300 -0.3479
#>  -0.2276  0.5595 -0.5377  0.4450  0.5906 -0.7494 -0.1824 -0.8327 -0.0951
#>   0.1774  0.0931 -0.3108 -0.3536  0.6550 -0.0635  0.3198 -0.6157 -0.4778
#> 
#> Columns 10 to 18 -0.6905 -0.2396 -0.7252 -0.7510 -0.2833 -0.6540  0.5175  0.2534  0.3038
#>  -0.7841 -0.3991  0.0567  0.2498 -0.7707 -0.3099 -0.4381  0.3531  0.1929
#>  -0.8555  0.2704  0.5437  0.2784 -0.7086 -0.7691  0.0976 -0.5864  0.2632
#> 
#> Columns 19 to 20  0.7931  0.0300
#>   0.4964 -0.7732
#>   0.5800  0.6527
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3082 -0.0897 -0.3844  0.1101 -0.0484  0.0566  0.3663 -0.2472 -0.2378
#>  -0.7401  0.4488 -0.6052 -0.3659  0.4683 -0.3227  0.1078 -0.0511 -0.2044
#>  -0.3908  0.0799  0.0532  0.0909  0.0916 -0.6912  0.3326  0.0922 -0.0194
#> 
#> Columns 10 to 18 -0.0648  0.0512 -0.2409 -0.0880 -0.5498 -0.5338  0.5173  0.0469 -0.8689
#>  -0.2667 -0.3246 -0.1317 -0.2257 -0.8091 -0.4201  0.1863  0.0253 -0.8763
#>  -0.2267 -0.5987 -0.4445 -0.1034 -0.1853 -0.0167  0.6349  0.5723 -0.0497
#> 
#> Columns 19 to 20  0.1100  0.3055
#>   0.3671 -0.5782
#>  -0.4821 -0.4685
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1596 -0.1483  0.0073  0.5861 -0.2487  0.0553  0.2920 -0.0950 -0.0575
#>  -0.1789 -0.0222 -0.2438  0.1613  0.3189 -0.3587  0.4551 -0.4676  0.5585
#>  -0.3287  0.2039 -0.6013  0.4982  0.3359 -0.6761  0.1696 -0.0010 -0.3536
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0587  0.2741 -0.1705 -0.7414  0.2654  0.5492 -0.3622  0.1313 -0.3726
#>  -0.5696  0.3732  0.5605 -0.2190 -0.0070  0.5764 -0.4673 -0.3701  0.0408
#>  -0.0861  0.3159 -0.3830 -0.2558  0.1244  0.4135 -0.2147  0.4287 -0.3692
#> 
#> Columns 10 to 18  0.4191  0.2423 -0.2241  0.3899  0.1271  0.3254  0.1230 -0.6684 -0.0752
#>  -0.5851 -0.4824  0.4365 -0.3866  0.3943  0.8684 -0.0551  0.2128 -0.2202
#>   0.0801  0.0747 -0.2568  0.5591  0.1612  0.5843 -0.2902 -0.1298 -0.3003
#> 
#> Columns 19 to 20  0.6735 -0.0037
#>   0.6704 -0.2653
#>   0.5996  0.1854
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0320  0.5045 -0.3379  0.4982  0.3064 -0.1904  0.0619 -0.0826 -0.2845
#>   0.0075  0.5180  0.1207 -0.2208 -0.0655 -0.0484  0.2749 -0.1938  0.2393
#>  -0.0091  0.4060 -0.1813  0.2407  0.1332  0.0128  0.3142 -0.0589  0.0608
#> 
#> Columns 10 to 18 -0.2352 -0.1703  0.2856  0.4529 -0.7087 -0.0851 -0.0944  0.6356 -0.5844
#>  -0.6021 -0.3737 -0.1312 -0.2531 -0.2873 -0.4735  0.1422  0.4830 -0.3397
#>  -0.2498 -0.4701 -0.0371  0.3842 -0.5420 -0.1499 -0.0703  0.2119 -0.3834
#> 
#> Columns 19 to 20 -0.3604 -0.3006
#>  -0.1569 -0.3843
#>   0.0581 -0.0047
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>