RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3348 -0.5385 -0.7527 -0.7822 -0.4131 -0.2305 -0.7594 -0.0388 -0.1475
#>   0.4695 -0.5969  0.0619  0.3989 -0.3510  0.5189 -0.1646  0.2304  0.0753
#>  -0.7290  0.3766  0.5027 -0.9391 -0.5970  0.4625  0.2705 -0.4882 -0.2239
#> 
#> Columns 10 to 18  0.4000 -0.8345  0.6953  0.7969  0.4153  0.2971  0.8391 -0.4422  0.4928
#>   0.6720  0.6425  0.3795  0.3424  0.6139  0.1998 -0.6441 -0.2749 -0.3978
#>   0.8533  0.6110 -0.7945  0.1752 -0.1129  0.6740  0.7455  0.7979  0.6226
#> 
#> Columns 19 to 20  0.7240 -0.2583
#>  -0.0410  0.7249
#>  -0.5640 -0.0470
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0442  0.0337  0.4087  0.4064  0.4453  0.1053  0.1310  0.3002  0.2253
#>  -0.2682 -0.5080 -0.8183  0.0267  0.1980 -0.0983  0.2186  0.2164 -0.1926
#>   0.3531 -0.5703 -0.2702  0.0072 -0.2709  0.5470 -0.1139  0.3808 -0.0705
#> 
#> Columns 10 to 18 -0.2842  0.4430 -0.1004  0.3836  0.6801  0.1613  0.6411 -0.8133  0.3003
#>   0.1223 -0.5581  0.1201 -0.1188  0.7002  0.2138  0.3841 -0.2292 -0.0751
#>  -0.2084  0.5637 -0.4440  0.1278  0.5048 -0.1715  0.4286 -0.0378 -0.1583
#> 
#> Columns 19 to 20  0.0138 -0.0352
#>   0.3657  0.0526
#>   0.0418 -0.3778
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3341 -0.2267 -0.2774 -0.5915  0.3036 -0.0643 -0.0624  0.1557 -0.3138
#>  -0.3216 -0.0424 -0.2451  0.1495  0.2673  0.1849  0.3201  0.3447 -0.0236
#>  -0.1212 -0.3303 -0.1545  0.4875  0.0470  0.1985 -0.6068  0.0853  0.4529
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8038 -0.0335  0.0576 -0.5498  0.2139  0.1737 -0.6327  0.1935  0.1120
#>  -0.4331 -0.1220  0.4650 -0.3498  0.5592  0.4342  0.4310 -0.4919  0.2385
#>   0.1220 -0.2756  0.8278 -0.4059 -0.1695  0.8619  0.4548 -0.3066  0.9076
#> 
#> Columns 10 to 18  0.1967 -0.9285  0.2272  0.0326  0.0167  0.6590 -0.8617 -0.7043  0.6058
#>  -0.8215 -0.5203  0.9065  0.2362  0.7014  0.0637  0.6052 -0.0426 -0.3881
#>  -0.5301 -0.7514  0.6590 -0.1785 -0.1746  0.7465  0.7957  0.5010  0.0897
#> 
#> Columns 19 to 20 -0.7364 -0.3728
#>   0.3344 -0.6150
#>   0.5250 -0.3850
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1425 -0.4643 -0.4756  0.5671  0.2961  0.5591 -0.0814  0.3130  0.0252
#>  -0.1243 -0.3274 -0.0844 -0.1061 -0.2465  0.1326 -0.5342 -0.0865 -0.3653
#>   0.1202 -0.3342 -0.3408 -0.2283  0.1756  0.5010 -0.3666  0.1527 -0.0161
#> 
#> Columns 10 to 18  0.0700  0.3731 -0.6963 -0.0513  0.2709  0.1697  0.2247 -0.4299  0.0194
#>   0.0855  0.3556 -0.1405  0.1742  0.3316  0.1293  0.5169 -0.3183  0.1045
#>   0.6171  0.5253 -0.6206 -0.0110  0.2648 -0.0406  0.5927 -0.4282  0.2079
#> 
#> Columns 19 to 20  0.3373  0.0761
#>   0.2659 -0.5187
#>   0.2070 -0.1419
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>