RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1756 -0.5571 -0.3656  0.2914  0.4896  0.4096  0.4326  0.6826 -0.1698
#>  -0.3826 -0.0843  0.7558  0.2018 -0.2105 -0.0060  0.3571 -0.3482 -0.2509
#>   0.3194 -0.7386  0.7120  0.5628  0.6523  0.3572 -0.6936  0.2913 -0.5284
#> 
#> Columns 10 to 18 -0.5237 -0.5278  0.6714 -0.7936  0.6469  0.7818 -0.1666  0.8763  0.3750
#>   0.0358  0.0858 -0.3561  0.3139  0.3720  0.7534 -0.5335  0.7458  0.3942
#>  -0.1901  0.3566 -0.4386  0.4104  0.2480  0.6405  0.3497  0.5617 -0.5879
#> 
#> Columns 19 to 20 -0.9159 -0.9268
#>  -0.8282 -0.8712
#>  -0.6471 -0.0348
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1523 -0.1414  0.2117 -0.5333 -0.6451 -0.6562  0.0767 -0.1539  0.0810
#>   0.5686 -0.4389  0.0764  0.0327  0.2468 -0.4043  0.0828  0.3111 -0.4268
#>   0.7847 -0.5039  0.8327  0.2129 -0.3419  0.0224  0.1604 -0.3103  0.3106
#> 
#> Columns 10 to 18  0.4389  0.1546  0.4886  0.2139  0.1144  0.2044 -0.2436 -0.2815  0.0738
#>  -0.3342 -0.0642  0.0661  0.4565 -0.0897 -0.3717  0.5649 -0.2800  0.5829
#>   0.0180  0.5505 -0.1313  0.0390  0.0171  0.0773 -0.6062 -0.0447  0.3274
#> 
#> Columns 19 to 20  0.6965  0.0408
#>   0.0885 -0.3700
#>   0.3598 -0.1298
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1307 -0.6029 -0.0983  0.3848  0.4506 -0.3311 -0.6324 -0.1000  0.5215
#>   0.3279 -0.2081  0.0752  0.4661  0.0012 -0.2298  0.1197 -0.3200  0.2454
#>   0.1415 -0.4636  0.0459  0.3436  0.2827  0.0557 -0.1511  0.3944  0.3037
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7775 -0.5286 -0.0298 -0.7098  0.6141 -0.1983  0.7386 -0.2353  0.0650
#>  -0.1815  0.0427  0.5183  0.1534 -0.0329  0.0160 -0.2392 -0.1283  0.3615
#>  -0.0756  0.1323  0.5043  0.0285  0.6020 -0.2258  0.1705 -0.1318 -0.1702
#> 
#> Columns 10 to 18  0.4348  0.3773 -0.1371 -0.4349  0.2777 -0.4116 -0.4484 -0.7135  0.3490
#>  -0.0197  0.1484 -0.4147 -0.1408  0.4613 -0.3726 -0.3722  0.4534  0.5735
#>   0.4189  0.0049 -0.2977 -0.1219  0.5140 -0.3259 -0.3805 -0.2714  0.5632
#> 
#> Columns 19 to 20  0.0260  0.0363
#>   0.2692 -0.0613
#>  -0.4531 -0.2075
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1364 -0.2502  0.4082  0.5785 -0.1171  0.0887 -0.0193  0.0385  0.6142
#>   0.5086 -0.4914  0.3308  0.1888  0.1076  0.0274 -0.1141  0.2595 -0.1287
#>   0.3920 -0.5201  0.6066  0.4702  0.1044  0.4700  0.2341 -0.1834  0.2802
#> 
#> Columns 10 to 18 -0.0925 -0.4908  0.5313  0.1662  0.0954  0.3769 -0.1071  0.3936 -0.2996
#>  -0.0545  0.0129 -0.0456  0.0184 -0.1397  0.3096 -0.2348  0.0390 -0.3388
#>  -0.1518  0.0787 -0.1228  0.1554 -0.0819  0.4687 -0.3007  0.1237 -0.2781
#> 
#> Columns 19 to 20 -0.4381 -0.1756
#>  -0.1244 -0.4948
#>  -0.2149 -0.3900
#> [ CPUFloatType{2,3,20} ]
#>