RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6288 -0.3388 -0.7945  0.2480 -0.6714 -0.4626 -0.7920 -0.8688  0.4303
#>  -0.2223 -0.3779  0.7331 -0.3861  0.5641  0.0288 -0.0222  0.4138 -0.5999
#>   0.0700  0.2211 -0.8866  0.6733 -0.5712  0.1803 -0.0205  0.6760 -0.6423
#> 
#> Columns 10 to 18  0.2244  0.1451 -0.7866  0.0980  0.1989  0.7395  0.5388  0.7991  0.4812
#>   0.3970 -0.6007  0.4471  0.0137 -0.1319  0.4433 -0.2709 -0.3821 -0.5953
#>   0.0437  0.4153 -0.6322 -0.9081 -0.8163 -0.6248  0.4869 -0.7317  0.4333
#> 
#> Columns 19 to 20 -0.0460  0.2984
#>   0.0382 -0.5404
#>  -0.4356  0.6416
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6418  0.1847  0.2821 -0.3665 -0.8397  0.4986 -0.2275 -0.4075 -0.7124
#>  -0.1558  0.3882  0.2318  0.5149 -0.6230  0.1155 -0.3507 -0.0645 -0.3351
#>   0.1606  0.0845 -0.1675 -0.6387 -0.5621  0.4172  0.6739  0.6678  0.0801
#> 
#> Columns 10 to 18  0.3332 -0.1502 -0.4022 -0.3053 -0.5556  0.1565  0.3363  0.3624  0.1940
#>   0.2896  0.0881 -0.3312 -0.5977 -0.0146  0.2858 -0.4104 -0.2720  0.1875
#>  -0.4854 -0.3105  0.0285  0.3492 -0.3603  0.0444  0.1789 -0.2588  0.2533
#> 
#> Columns 19 to 20 -0.5943 -0.2363
#>  -0.2176 -0.4102
#>  -0.4662  0.0549
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1657 -0.0795  0.4598 -0.4607 -0.5324  0.2152 -0.3294 -0.3832 -0.3537
#>   0.1237  0.4820  0.1751 -0.5140 -0.6225  0.2608  0.1560  0.0662 -0.0944
#>   0.1200 -0.1709  0.3212  0.2351 -0.1041 -0.1697  0.1233  0.3035  0.0389
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6107  0.4707 -0.8053 -0.1884  0.1207 -0.1824  0.2034  0.6101  0.4930
#>   0.0770 -0.0275 -0.1535 -0.2858  0.6341 -0.1619  0.3811 -0.1302  0.5874
#>  -0.1801  0.1689  0.2646 -0.0433  0.3275 -0.2273  0.4608  0.1591  0.5854
#> 
#> Columns 10 to 18  0.2670 -0.3498 -0.1070 -0.1099  0.5448  0.7573  0.0192  0.5470  0.4062
#>   0.5496  0.1264 -0.4316  0.2321  0.6703  0.2167  0.1607  0.2692 -0.6893
#>   0.1168  0.7897  0.1187 -0.0552  0.5289  0.0401 -0.1588  0.6160 -0.2593
#> 
#> Columns 19 to 20 -0.0465 -0.5472
#>  -0.1806 -0.1686
#>  -0.2058  0.3701
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2992  0.5366  0.3290 -0.4236 -0.2271  0.2681 -0.0582  0.1981 -0.2935
#>   0.4174  0.2989  0.2699 -0.3500 -0.3749  0.3494 -0.1921  0.0838 -0.0365
#>   0.2041  0.2039  0.0787 -0.4782 -0.7080  0.3227 -0.0161 -0.2870 -0.3469
#> 
#> Columns 10 to 18  0.1615  0.5327 -0.4477  0.2190 -0.1402 -0.0436  0.0977 -0.5480  0.2310
#>   0.1275  0.3665 -0.2063  0.0730 -0.3976  0.0309  0.1981 -0.0020  0.3022
#>  -0.2730  0.2032 -0.2453  0.2314 -0.4935 -0.3786  0.2959 -0.1024  0.1559
#> 
#> Columns 19 to 20 -0.3100  0.1452
#>  -0.1605  0.3437
#>  -0.2695 -0.2396
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>