RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0537 -0.6068  0.8033  0.3967  0.0185  0.2032 -0.6807 -0.7084 -0.3582
#>  -0.4092  0.4842  0.2756  0.4547 -0.1914 -0.7162 -0.0336  0.0121  0.0648
#>  -0.3972  0.3814  0.7463  0.7204  0.7124  0.1215 -0.0898 -0.4505 -0.6863
#> 
#> Columns 10 to 18 -0.8639 -0.5986 -0.7927  0.6161  0.5645  0.1213 -0.8038  0.6145 -0.2506
#>   0.2186  0.4810  0.3862 -0.2910 -0.1779 -0.2470  0.1085  0.6626 -0.2300
#>   0.4461 -0.0846 -0.2960 -0.1374  0.2031 -0.4429 -0.8438 -0.7429  0.2023
#> 
#> Columns 19 to 20 -0.0342  0.0967
#>   0.6556  0.2216
#>   0.4869 -0.9280
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1810  0.4723 -0.2617  0.1098 -0.2673  0.4500 -0.3125  0.4164  0.0894
#>   0.4749 -0.0708 -0.0366  0.0410  0.2128  0.1320 -0.3991 -0.0076 -0.3653
#>   0.4374  0.1538  0.1433 -0.0633  0.1126 -0.0567 -0.3892 -0.1950 -0.4902
#> 
#> Columns 10 to 18  0.3574  0.6086  0.0286 -0.3772  0.3669 -0.0733  0.3387  0.0259 -0.0769
#>   0.0719 -0.0935  0.4362  0.0645 -0.0592 -0.3246 -0.4009 -0.1299  0.1822
#>  -0.3642  0.4127 -0.1603  0.2333  0.0077  0.1191 -0.0255  0.4734  0.0880
#> 
#> Columns 19 to 20  0.1983  0.1424
#>   0.1864  0.2763
#>   0.5885 -0.1996
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0553 -0.1301  0.6641 -0.1184  0.3832  0.0291 -0.2848 -0.4397  0.1988
#>   0.2723 -0.2165  0.4338  0.2240 -0.1792  0.2593 -0.1474 -0.2377 -0.0933
#>  -0.0487 -0.3329  0.1972 -0.3834  0.0012  0.2285 -0.5744 -0.4609 -0.2945
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1068 -0.3323  0.1134  0.5120 -0.4304 -0.6575  0.2735  0.0121  0.1635
#>   0.3330 -0.3108  0.0146  0.3557  0.7578 -0.3226 -0.4828 -0.3055 -0.0002
#>   0.0043  0.2445 -0.2699  0.4356 -0.7314 -0.7342  0.4853  0.2596 -0.4064
#> 
#> Columns 10 to 18  0.5720  0.0313 -0.3734 -0.0175  0.1412  0.5876 -0.6988 -0.1491  0.6670
#>   0.2865  0.7203 -0.3529  0.2108 -0.1786 -0.0720 -0.4750 -0.1662  0.4783
#>   0.5502  0.5487  0.2137 -0.1094 -0.0502  0.4483  0.1876  0.3796 -0.1206
#> 
#> Columns 19 to 20 -0.2593  0.0640
#>  -0.2874  0.2052
#>   0.4311  0.0880
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1197 -0.1857  0.5315  0.1047  0.2196  0.2221 -0.3519  0.0146  0.0937
#>   0.0829  0.0967  0.4527  0.3903  0.1215  0.3224 -0.0493 -0.3118 -0.0241
#>  -0.2004  0.0826 -0.0198  0.3878 -0.4106  0.3690 -0.1828  0.1567  0.1521
#> 
#> Columns 10 to 18  0.2606  0.2827  0.2827 -0.0838 -0.0542 -0.3670  0.1961  0.0280 -0.1036
#>   0.1262 -0.1101  0.0760  0.1585 -0.0776 -0.2045 -0.2221  0.1665 -0.0193
#>   0.2738  0.0870  0.3653 -0.1080  0.3085 -0.1807 -0.1262  0.2102 -0.4338
#> 
#> Columns 19 to 20  0.0575 -0.0439
#>   0.3363  0.2630
#>   0.1710 -0.6024
#> [ CPUFloatType{2,3,20} ]
#>