RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6252 -0.7812  0.5629  0.1690  0.1904 -0.8417 -0.0147  0.6596  0.2426
#>   0.6824 -0.3727 -0.6970 -0.0224 -0.1399 -0.4292 -0.6254 -0.3455 -0.1364
#>   0.0077 -0.0550  0.9061  0.2386  0.3614 -0.6421 -0.3014  0.2091  0.5724
#> 
#> Columns 10 to 18 -0.3552  0.5974 -0.5374 -0.0052  0.4183  0.2892  0.1245  0.3733 -0.6295
#>  -0.5138  0.0457 -0.2365  0.5713  0.1856  0.8353 -0.6914 -0.7508  0.4005
#>  -0.2340  0.6952 -0.3689  0.0510  0.1070  0.1017 -0.6752 -0.7684 -0.1997
#> 
#> Columns 19 to 20  0.6894 -0.1300
#>   0.2744 -0.1648
#>   0.6123 -0.5418
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5380 -0.0012  0.4703 -0.3088  0.0730 -0.2425  0.0880 -0.0256  0.2412
#>   0.3343 -0.1167  0.2327  0.0229  0.4965 -0.6581  0.0770  0.5828  0.1755
#>  -0.4168  0.6234  0.3170 -0.0886  0.2885  0.1450  0.2653  0.2405  0.2914
#> 
#> Columns 10 to 18  0.3198  0.2569 -0.0403 -0.4338  0.2494 -0.3516  0.1421  0.1031  0.2720
#>  -0.2485  0.6235 -0.1719 -0.2492 -0.2467  0.1399 -0.2427  0.4591 -0.2121
#>  -0.3484  0.0279 -0.1515 -0.3776  0.2098 -0.4960  0.3716  0.3217 -0.6030
#> 
#> Columns 19 to 20  0.7011  0.0889
#>   0.1224  0.0803
#>   0.1795 -0.4655
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0839  0.1023  0.6420 -0.2826 -0.2290  0.0888  0.4524 -0.0426 -0.0564
#>   0.5510 -0.2067  0.2664 -0.4532  0.2875 -0.1704 -0.3782  0.5942  0.3296
#>  -0.1719  0.1359  0.6681 -0.1231 -0.4445  0.1658  0.4587  0.2792 -0.3222
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1903  0.2250  0.0962 -0.0251 -0.2078  0.0334 -0.0570  0.2278  0.1578
#>  -0.2804 -0.6637  0.3607  0.3279  0.2614  0.9112  0.3800  0.1949 -0.2424
#>   0.2766 -0.0872  0.1213  0.2005  0.4192  0.3856  0.6156 -0.2634 -0.0630
#> 
#> Columns 10 to 18  0.0196  0.0987  0.0527 -0.1394  0.2361  0.3517 -0.3229 -0.0434  0.3661
#>  -0.1270  0.4994  0.3451 -0.6262  0.2084 -0.1199 -0.1820 -0.3744  0.4413
#>  -0.1290  0.4806 -0.0676 -0.4265 -0.1269  0.5368  0.7728  0.1658  0.3623
#> 
#> Columns 19 to 20  0.1843  0.0385
#>   0.6115 -0.1134
#>  -0.5248 -0.1324
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1781  0.0368  0.3126 -0.3617 -0.3079 -0.1915  0.1792  0.4797 -0.0725
#>  -0.0038 -0.2524  0.7926 -0.1418 -0.0712  0.0549  0.7449  0.4675  0.3382
#>   0.0659  0.2423  0.6511 -0.3877 -0.0205  0.0546  0.4681  0.2179  0.1197
#> 
#> Columns 10 to 18 -0.1500  0.3876  0.0313 -0.1525  0.1141 -0.0498  0.2144 -0.3037  0.0541
#>  -0.3634  0.2878 -0.2373 -0.4750  0.4619 -0.3576  0.2476  0.0022 -0.5040
#>  -0.0171  0.1932  0.1014 -0.3744  0.2911 -0.4318  0.0421 -0.0505 -0.0474
#> 
#> Columns 19 to 20  0.5708 -0.2972
#>   0.3489 -0.5191
#>   0.5218 -0.3960
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>