RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3173 -0.8960  0.5295 -0.6393 -0.0437 -0.6835  0.2393  0.5844  0.0967
#>   0.7159  0.0327 -0.5300 -0.1070 -0.0217  0.3576 -0.2554  0.0789  0.7892
#>  -0.5832  0.2636  0.6732  0.8859  0.5352 -0.3255  0.4616 -0.7095 -0.7352
#> 
#> Columns 10 to 18 -0.0989 -0.8542 -0.8077  0.0632  0.2810  0.8634 -0.7744 -0.5660 -0.1092
#>  -0.8761  0.2234 -0.4614 -0.5741 -0.9090  0.6398 -0.2125 -0.7536  0.1970
#>   0.4314  0.3749 -0.3353 -0.3697 -0.4498  0.1021  0.7234 -0.3354 -0.3148
#> 
#> Columns 19 to 20  0.7228  0.3498
#>   0.1839  0.0696
#>  -0.0750 -0.7186
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1009 -0.6942  0.6321  0.1955  0.3233  0.3183  0.5272  0.2457 -0.6240
#>   0.2604  0.2015 -0.2045  0.0065 -0.2419  0.2768 -0.1866  0.4592 -0.1899
#>   0.6623 -0.1296  0.1506  0.4710 -0.0543  0.4035  0.0173  0.6203 -0.5398
#> 
#> Columns 10 to 18  0.5028  0.2904  0.7242  0.6015  0.6863 -0.5025 -0.0031 -0.0433  0.3524
#>   0.7331 -0.1786 -0.0203 -0.0817  0.1677 -0.3659 -0.4257 -0.0586  0.5066
#>   0.1416 -0.4348  0.6351  0.2326 -0.1898 -0.2155 -0.4175 -0.0603  0.3497
#> 
#> Columns 19 to 20  0.1495  0.0646
#>  -0.0951 -0.0371
#>   0.6244 -0.1887
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3861 -0.0519 -0.2524 -0.1258 -0.2608 -0.0131 -0.0288  0.2863 -0.0363
#>   0.1458  0.0431 -0.0529 -0.0128 -0.2555  0.0267 -0.1477 -0.1088 -0.0069
#>   0.3224  0.0549 -0.0788 -0.1045 -0.7161 -0.0246  0.2003 -0.0943  0.1891
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6487 -0.2268 -0.5649  0.0559  0.1883 -0.0694  0.5136 -0.3853 -0.1155
#>   0.5737 -0.2090 -0.5212  0.1609 -0.0766  0.3984  0.0273 -0.4492  0.0409
#>   0.3949  0.6716 -0.6980 -0.4705 -0.2080  0.2473  0.6799 -0.2593  0.2797
#> 
#> Columns 10 to 18 -0.4799 -0.5476 -0.6190  0.2539 -0.2861 -0.8043 -0.1753 -0.4099 -0.4034
#>   0.3052  0.0869  0.3730 -0.1295 -0.2095  0.3074 -0.7868 -0.0503  0.3129
#>  -0.1571 -0.4666  0.6874  0.4876  0.2015  0.1216 -0.1143 -0.6277 -0.6153
#> 
#> Columns 19 to 20  0.5876  0.3658
#>   0.0188  0.5181
#>  -0.2409 -0.1798
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0367 -0.0927  0.1210 -0.3147 -0.5949 -0.1655 -0.0024  0.0078  0.0244
#>   0.0798 -0.2214  0.0933 -0.2382 -0.0852  0.1978  0.0132  0.2341 -0.2896
#>   0.5926 -0.2339  0.2549 -0.3987 -0.4111  0.3076  0.0941  0.4885 -0.6499
#> 
#> Columns 10 to 18  0.4735 -0.3362  0.0944  0.5118 -0.1879  0.2169 -0.2219  0.0838 -0.2780
#>   0.1384 -0.4200 -0.0089  0.5506  0.4923 -0.2645  0.0014  0.1860 -0.3958
#>   0.3557 -0.3081  0.0193  0.2306  0.3715 -0.3026 -0.3790 -0.1606 -0.0093
#> 
#> Columns 19 to 20  0.3700  0.0929
#>   0.2160  0.0375
#>  -0.2639  0.4242
#> [ CPUFloatType{2,3,20} ]
#>