RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2676 -0.3452 -0.7077 -0.0424 -0.0504  0.1090  0.6053 -0.2870  0.5964
#>  -0.1282 -0.8201 -0.7097  0.2440 -0.1272 -0.2654  0.2774  0.3185  0.4276
#>   0.7312 -0.2406 -0.1339 -0.0869 -0.3075 -0.6807 -0.1037 -0.8999  0.1945
#> 
#> Columns 10 to 18  0.0266 -0.0101  0.0047  0.0895  0.0382  0.1314  0.4702  0.2849 -0.7895
#>   0.7020  0.8665 -0.3360  0.1030  0.3367 -0.2105 -0.8239 -0.2160  0.0424
#>  -0.0688 -0.2532  0.2012  0.6533 -0.7055 -0.2762  0.2882  0.9247 -0.8194
#> 
#> Columns 19 to 20  0.1414  0.8436
#>  -0.3012 -0.2645
#>   0.6284  0.5333
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3390  0.3559  0.6169 -0.4054  0.3527  0.1210  0.1920  0.3939  0.0031
#>   0.4944  0.1728  0.0099  0.1891  0.3769  0.4861 -0.1781  0.0492  0.3818
#>   0.5340 -0.5014  0.1113 -0.6090  0.7400  0.3519  0.2818  0.0885  0.1197
#> 
#> Columns 10 to 18 -0.4586 -0.5061 -0.0555  0.4286 -0.6161 -0.2624  0.0823 -0.3447 -0.1342
#>  -0.3090  0.2611  0.3481  0.1499 -0.4297 -0.2957 -0.0033 -0.3452 -0.1002
#>   0.0306 -0.8030 -0.6825  0.7034 -0.7695 -0.5208 -0.1887 -0.1813 -0.3442
#> 
#> Columns 19 to 20  0.2611  0.7128
#>  -0.3192  0.4823
#>   0.6100 -0.0965
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1282 -0.3302  0.5616  0.0262  0.6103  0.0166  0.0729 -0.3328 -0.6775
#>   0.1695 -0.6144 -0.3247 -0.3754  0.4078  0.0503  0.3815  0.3532  0.1950
#>   0.2500  0.1938 -0.2983  0.0609  0.4863  0.0396  0.1087  0.5329 -0.3151
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0583  0.7644 -0.2088  0.3974  0.3604 -0.6307 -0.2311 -0.1932 -0.2396
#>  -0.4051 -0.1978  0.3018 -0.3689  0.4720 -0.3011 -0.4334 -0.1600  0.5369
#>  -0.7167  0.3906  0.5215  0.1094 -0.0373 -0.3576 -0.0714  0.1790  0.3556
#> 
#> Columns 10 to 18 -0.3424  0.4203  0.1997 -0.1287  0.5289 -0.2457 -0.0831 -0.0260 -0.4322
#>  -0.4600 -0.3096 -0.8875 -0.2274  0.0858  0.1538  0.0638 -0.2072  0.0586
#>   0.4392 -0.7636 -0.2142  0.7072 -0.3693  0.4618 -0.3212  0.5427  0.0710
#> 
#> Columns 19 to 20 -0.2828  0.7044
#>   0.0533 -0.3841
#>   0.1850 -0.3588
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3295  0.3244  0.0652 -0.0411  0.4235  0.6687  0.0323  0.7192 -0.2114
#>  -0.0825 -0.2318  0.0484  0.1029  0.5272  0.2252  0.2108 -0.4102 -0.4723
#>   0.2576 -0.2356  0.3224 -0.4624  0.6218  0.1679  0.1245  0.2489 -0.3357
#> 
#> Columns 10 to 18 -0.6234 -0.0558  0.2208  0.1468 -0.5174 -0.0961 -0.3960  0.0632  0.1961
#>  -0.3363 -0.0498  0.0160  0.3882 -0.3920 -0.0349 -0.0433  0.1103 -0.4033
#>  -0.1472 -0.2943 -0.2443  0.3658 -0.5304  0.2259 -0.2703 -0.3997 -0.6190
#> 
#> Columns 19 to 20 -0.1037  0.5050
#>   0.3125  0.6547
#>   0.2745  0.1527
#> [ CPUFloatType{2,3,20} ]
#>