RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7661  0.6920 -0.8804 -0.6727 -0.3631 -0.7018 -0.3121 -0.7370  0.1601
#>   0.8320  0.4168  0.3258 -0.1105  0.3153  0.1602  0.5944 -0.4153 -0.9204
#>  -0.7091  0.7673  0.4333  0.8686  0.3759 -0.0716  0.4846 -0.4814 -0.5964
#> 
#> Columns 10 to 18  0.9547  0.2620 -0.4947  0.8918 -0.6644  0.7605 -0.2719  0.9166  0.0472
#>   0.8223 -0.5200 -0.9433 -0.5360  0.8119  0.7434  0.0888  0.7802 -0.7634
#>   0.2602  0.6097 -0.0188  0.5509 -0.1346 -0.4331 -0.7859 -0.0353 -0.5925
#> 
#> Columns 19 to 20 -0.0298  0.3819
#>   0.3262 -0.5102
#>   0.5254 -0.5418
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5715  0.3147  0.7375  0.7698  0.2941 -0.4362 -0.2013 -0.1352 -0.7984
#>  -0.8553 -0.5224  0.6650  0.7063 -0.4008 -0.7937 -0.0702  0.0546  0.0660
#>   0.0897 -0.4419  0.0483 -0.4195  0.2427 -0.1832 -0.4332  0.2983 -0.6891
#> 
#> Columns 10 to 18 -0.0835  0.4937 -0.3261  0.0866  0.3088  0.2473 -0.6166  0.4145 -0.2415
#>   0.2450  0.4723  0.2515  0.7859 -0.3040 -0.0121 -0.3335 -0.6250 -0.1210
#>   0.2420 -0.3401 -0.4045  0.4137  0.0103  0.6891 -0.5636  0.3785 -0.4333
#> 
#> Columns 19 to 20  0.0156 -0.5718
#>  -0.8111 -0.7515
#>  -0.1365 -0.6682
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.8165 -0.1558  0.3512  0.3217 -0.3637 -0.5244 -0.1538  0.3354 -0.1854
#>  -0.3006  0.3180 -0.3176 -0.3453  0.1891 -0.6390 -0.6548  0.0073 -0.5005
#>  -0.7851  0.5488  0.2025  0.5676 -0.4726 -0.7423 -0.3410  0.1968  0.1106
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0000 -0.1061 -0.1191  0.7814  0.2315 -0.2851  0.6424 -0.3096 -0.8192
#>   0.3166  0.3930  0.0022 -0.4133  0.1140  0.3509 -0.0437  0.2747  0.3911
#>  -0.1530  0.3723  0.0236 -0.2888  0.1906  0.2836  0.1568 -0.6986  0.1518
#> 
#> Columns 10 to 18  0.0411  0.0464  0.1280  0.5439 -0.2654 -0.7867  0.8149  0.1430  0.5036
#>  -0.0600 -0.0411  0.6346  0.0426 -0.0289 -0.7022  0.3128 -0.3454 -0.5493
#>   0.5540  0.0301 -0.0657 -0.0832  0.4594  0.2420  0.3184 -0.1955  0.1230
#> 
#> Columns 19 to 20 -0.8061 -0.6732
#>  -0.8462 -0.6112
#>   0.2414 -0.2440
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6941  0.4185  0.1213  0.5897 -0.3819 -0.6368 -0.1015  0.4235 -0.3774
#>  -0.7026 -0.1828 -0.1087  0.1038  0.1645 -0.3736 -0.2081  0.4129 -0.7072
#>  -0.5302  0.0627  0.0791  0.5525  0.0848 -0.2454 -0.2109  0.2500 -0.1381
#> 
#> Columns 10 to 18 -0.0010 -0.4334  0.2498  0.1653  0.2920  0.3646 -0.3384  0.4474  0.0907
#>  -0.3470  0.0917 -0.1274 -0.1704 -0.6042  0.4078 -0.0561 -0.1272 -0.1464
#>  -0.2483 -0.0683 -0.0269  0.2745 -0.4832  0.6082 -0.3021  0.0662  0.0733
#> 
#> Columns 19 to 20 -0.5503 -0.2289
#>  -0.1400 -0.1379
#>   0.1213 -0.0651
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>