RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3073 -0.5243 -0.2121  0.1086  0.5067  0.1896  0.7125 -0.7447 -0.4688
#>  -0.5717 -0.2075  0.2333  0.4458  0.1447 -0.5718  0.5406  0.5820  0.2825
#>  -0.0570 -0.1874  0.3118 -0.1447  0.6547  0.7762  0.1354 -0.1181  0.3002
#> 
#> Columns 10 to 18  0.1306 -0.6932  0.0634 -0.3089  0.0434 -0.7083  0.9037  0.8700  0.3639
#>  -0.7720  0.5384  0.2505 -0.4550 -0.7595  0.3701  0.0437 -0.3190  0.7274
#>  -0.5555 -0.1185  0.2661 -0.4051 -0.2690 -0.0595 -0.5976 -0.7418  0.7395
#> 
#> Columns 19 to 20 -0.4230  0.0184
#>   0.8457  0.1715
#>   0.8040  0.3551
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3583  0.4417  0.3369 -0.1763 -0.4919  0.3386 -0.1548 -0.6026 -0.2792
#>  -0.0439  0.0667 -0.5315  0.5501  0.3661 -0.0437  0.6350 -0.0089 -0.2472
#>  -0.3925  0.3833  0.2045  0.1230  0.3654 -0.0310  0.3891  0.1703 -0.0397
#> 
#> Columns 10 to 18  0.0355  0.0189  0.3944 -0.3036 -0.0006 -0.3663 -0.3670 -0.3137 -0.0321
#>   0.0312  0.1141  0.5869 -0.3732  0.0388  0.2023  0.3577 -0.0037  0.3621
#>   0.3534  0.2503  0.3357 -0.0461  0.2389 -0.2225 -0.2133 -0.4707 -0.6483
#> 
#> Columns 19 to 20  0.2254  0.0025
#>   0.1950  0.2727
#>   0.5743  0.5588
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.6939  0.4674  0.5693  0.5794  0.2155  0.1815  0.4945 -0.3649 -0.2034
#>  -0.2023  0.5637 -0.2841  0.3501  0.3116 -0.0196  0.3686 -0.1085 -0.1109
#>  -0.0868  0.5847 -0.3372  0.3520  0.6570 -0.6093  0.8438  0.0601  0.0004
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0814  0.6798 -0.4152 -0.1398 -0.2256  0.6268  0.0044 -0.2954 -0.2566
#>  -0.5932  0.4592  0.3974 -0.4657  0.6193  0.2562 -0.6941 -0.0074  0.3169
#>  -0.6281  0.2885  0.2360 -0.5721 -0.2235 -0.4707 -0.6370 -0.0182  0.3130
#> 
#> Columns 10 to 18 -0.5390 -0.2732  0.1272 -0.6677 -0.4859  0.1404  0.2967 -0.4652  0.0146
#>   0.1670  0.2682 -0.3068  0.1586  0.2896  0.2624  0.3597  0.1550  0.0218
#>  -0.7229  0.7148 -0.5403  0.7543 -0.0362  0.0318  0.3935  0.1557 -0.4845
#> 
#> Columns 19 to 20  0.2567 -0.7705
#>  -0.1740 -0.6036
#>  -0.1394 -0.4611
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5294  0.6642 -0.3314  0.1181  0.2937  0.2790  0.7106 -0.0951 -0.5280
#>  -0.2011  0.6628 -0.2325  0.0260  0.5380 -0.3194  0.6148 -0.0966 -0.3220
#>  -0.1903  0.6110  0.2156  0.1509  0.4289 -0.2166  0.8068 -0.4552 -0.3280
#> 
#> Columns 10 to 18  0.1931 -0.3516  0.2777 -0.3735  0.2647 -0.5747  0.4943  0.3255  0.3662
#>   0.4062  0.2386 -0.2772  0.0431 -0.0539 -0.5318  0.0143  0.1554 -0.3728
#>   0.3916  0.2244  0.0176  0.0806 -0.3187 -0.5493  0.1354  0.4246  0.1651
#> 
#> Columns 19 to 20  0.1392  0.2905
#>   0.5494  0.2433
#>   0.5399  0.3858
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>