RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7372 -0.3097  0.3437  0.1108  0.6923  0.3836  0.0014  0.8093  0.0532
#>   0.3673 -0.2752  0.3866  0.0535 -0.2212 -0.8820 -0.8721  0.0650  0.3748
#>   0.1176  0.7874 -0.8794  0.9177 -0.7249 -0.6650 -0.1280 -0.0407 -0.1657
#> 
#> Columns 10 to 18  0.0336  0.0070  0.2253  0.5923 -0.7624  0.2862 -0.2399 -0.6879  0.2558
#>  -0.6848 -0.6200 -0.1119  0.8183  0.4157 -0.8426  0.3020  0.0024  0.0244
#>   0.6601  0.3522  0.2696 -0.0379  0.0943  0.2091  0.5489  0.5329  0.6434
#> 
#> Columns 19 to 20 -0.7001  0.5318
#>   0.8434 -0.5430
#>   0.3113  0.7515
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1053 -0.4250  0.5410  0.4200 -0.4588 -0.5663 -0.1886  0.0398  0.0936
#>  -0.2849 -0.1970  0.1411 -0.0003 -0.3104 -0.7767 -0.4322 -0.0048  0.2382
#>  -0.0984 -0.0862 -0.1531  0.2567 -0.1461 -0.0689 -0.0982 -0.2289 -0.1553
#> 
#> Columns 10 to 18  0.0289 -0.3094  0.2200  0.4417  0.5223 -0.2539 -0.4630  0.3220  0.1733
#>   0.1571  0.3873  0.1428  0.2050 -0.2783 -0.0307  0.0553  0.1290  0.3275
#>   0.6216 -0.2622  0.1002 -0.3186 -0.7407  0.0688  0.5578  0.6760  0.2099
#> 
#> Columns 19 to 20  0.1040  0.1753
#>  -0.5678 -0.2374
#>   0.0168 -0.3000
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5214  0.1367  0.2317  0.2635 -0.5539 -0.6253 -0.1277  0.3825  0.2454
#>   0.6708  0.1590 -0.3448  0.0153  0.0477  0.0873 -0.2882  0.1926 -0.1309
#>   0.3009 -0.4320  0.2499  0.0632 -0.6711 -0.3430  0.0169  0.0775  0.2092
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7200 -0.2387 -0.7245 -0.5723 -0.7732  0.4799 -0.4030  0.9101 -0.4457
#>   0.3799 -0.0809  0.7332 -0.0831 -0.2977  0.5339  0.5231  0.5631 -0.6382
#>   0.8773  0.3451  0.2036  0.4354 -0.5239  0.1748 -0.5900  0.5450  0.0829
#> 
#> Columns 10 to 18  0.3291  0.5033 -0.2655  0.6193 -0.0993 -0.7336  0.6966  0.3673 -0.0585
#>   0.5329 -0.1903 -0.5771 -0.3730  0.3111  0.1254  0.8493  0.7299  0.4403
#>   0.1577  0.6610  0.0072 -0.0728 -0.6491 -0.7580  0.1878 -0.4176  0.2215
#> 
#> Columns 19 to 20  0.4701  0.6285
#>   0.2686  0.5301
#>   0.8902  0.0387
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3771 -0.2379 -0.0733 -0.4137 -0.0207  0.5901 -0.1714  0.0023 -0.1333
#>   0.0573 -0.4406  0.0981 -0.5059  0.0957 -0.0609 -0.2763  0.0080  0.4065
#>  -0.3939 -0.1945 -0.1091 -0.1454 -0.4816 -0.1633  0.0673  0.2593 -0.4532
#> 
#> Columns 10 to 18 -0.0104 -0.2346 -0.4190 -0.5262 -0.5457  0.5673  0.4492  0.0148 -0.2816
#>  -0.1136 -0.2903  0.1355 -0.3324 -0.1291  0.0433  0.1454  0.4710 -0.4753
#>   0.4407  0.5837 -0.0057 -0.2216 -0.1795  0.3427  0.3423  0.2905  0.4225
#> 
#> Columns 19 to 20  0.1037 -0.4711
#>   0.2862 -0.2633
#>  -0.1808 -0.1211
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>