RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1716  0.2015  0.3022 -0.6722 -0.0226  0.0535  0.6816 -0.1172  0.7797
#>   0.7180 -0.6351  0.2647  0.5633  0.4399  0.5392  0.4484  0.6734 -0.5390
#>   0.0311  0.3223 -0.0942 -0.8152  0.7113 -0.2346  0.0843  0.0440  0.2244
#> 
#> Columns 10 to 18  0.7318  0.2736  0.3619  0.7510 -0.0095 -0.0550  0.5170  0.6410  0.4223
#>  -0.4053 -0.6575 -0.4560 -0.4068 -0.1159 -0.5185  0.1890  0.3108 -0.7712
#>   0.0834 -0.6319 -0.5534 -0.6101 -0.2521 -0.7169  0.6075 -0.1566 -0.7217
#> 
#> Columns 19 to 20 -0.4590 -0.8597
#>  -0.7176 -0.6101
#>  -0.1733 -0.5453
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6301  0.2190  0.7769  0.0496  0.1941 -0.1590 -0.6686 -0.0251 -0.1371
#>   0.5423 -0.4392 -0.5229 -0.5453 -0.6948  0.4804 -0.0505  0.2858 -0.6719
#>  -0.3205  0.5507  0.2472 -0.4196 -0.2852  0.0970 -0.1025  0.4459 -0.3264
#> 
#> Columns 10 to 18  0.5090 -0.4465 -0.0750 -0.1161 -0.2479 -0.0879 -0.8236 -0.1428 -0.2302
#>   0.1618  0.0802  0.8672 -0.2703  0.1410  0.2639 -0.2427 -0.1479 -0.7456
#>  -0.1164 -0.4204  0.3176 -0.0622 -0.4702  0.2008 -0.1326 -0.7370 -0.1115
#> 
#> Columns 19 to 20  0.4341  0.3306
#>   0.3902  0.0656
#>   0.4307 -0.1582
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0487  0.1221  0.0852 -0.4624 -0.1770  0.4038  0.2900  0.2132 -0.2770
#>   0.1715  0.3170  0.2702 -0.2617  0.1906  0.1412  0.2035  0.2174 -0.1032
#>   0.1835  0.2381  0.1213 -0.1211  0.1174 -0.0434  0.4653  0.5618 -0.3993
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1774  0.2533 -0.7257 -0.0205  0.3437  0.4907 -0.2900  0.2619  0.1032
#>  -0.8303 -0.2361  0.0144 -0.0552 -0.4645 -0.6331  0.2586  0.3161 -0.0691
#>  -0.1842  0.2359 -0.1658  0.4742 -0.0792  0.7925  0.5630  0.7174 -0.1038
#> 
#> Columns 10 to 18  0.5769  0.4369 -0.2006  0.1633 -0.3756 -0.0390  0.0933 -0.4977  0.1441
#>  -0.1291  0.5214 -0.7887 -0.6792  0.1210 -0.2323  0.8151 -0.0213  0.5709
#>   0.2293 -0.0774 -0.5107 -0.2018 -0.1184 -0.3659 -0.4516  0.0288  0.4432
#> 
#> Columns 19 to 20  0.0943  0.7943
#>   0.1235  0.5337
#>  -0.0738  0.2081
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1692  0.0263  0.4274 -0.1926  0.0967 -0.1416 -0.3307  0.0412 -0.2719
#>  -0.2114  0.2464 -0.0489 -0.3525 -0.0045 -0.1038  0.4405  0.4135 -0.2294
#>   0.0762 -0.2973 -0.0367 -0.4062 -0.2226  0.1862  0.2824 -0.0989 -0.3454
#> 
#> Columns 10 to 18  0.2107 -0.4414 -0.1635 -0.6444 -0.0133 -0.1030 -0.1853 -0.4472 -0.2757
#>   0.1591 -0.5850 -0.2912 -0.2261 -0.3599 -0.3566  0.1753 -0.4643 -0.0657
#>   0.0788 -0.0253 -0.5884 -0.1894 -0.0457  0.0434  0.1939 -0.0162  0.0949
#> 
#> Columns 19 to 20 -0.3450 -0.1131
#>  -0.3640 -0.0334
#>  -0.1054  0.2523
#> [ CPUFloatType{2,3,20} ]
#>