RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7838 -0.2724  0.2219  0.8535  0.3488 -0.9109 -0.8632  0.8879 -0.0430
#>   0.1194  0.8681  0.8039  0.7592  0.6184  0.5210 -0.0173  0.8240 -0.6162
#>   0.8149 -0.6358 -0.3997 -0.7019  0.1908  0.0210  0.9786 -0.4196  0.7319
#> 
#> Columns 10 to 18  0.8718 -0.8775  0.0757  0.8592  0.5724  0.6404 -0.1646 -0.4236  0.6125
#>   0.2819 -0.4282 -0.3586 -0.3911  0.5005  0.5080 -0.4780  0.5320 -0.3374
#>   0.1253  0.6637  0.2802  0.6579 -0.0746 -0.1939  0.3719 -0.2410  0.4824
#> 
#> Columns 19 to 20  0.8898 -0.1842
#>   0.4000  0.7079
#>   0.5544  0.3637
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7196 -0.4827 -0.3589  0.2095 -0.1374 -0.1370 -0.2742  0.0122 -0.2104
#>  -0.4836 -0.6810 -0.6549  0.3556 -0.0903 -0.2255 -0.2713  0.2195  0.1796
#>  -0.4024 -0.0052  0.4738 -0.4304 -0.0789 -0.3927 -0.1739 -0.7033  0.5778
#> 
#> Columns 10 to 18  0.4356 -0.7240 -0.3272  0.0704 -0.4845  0.2215  0.0805  0.5582  0.0806
#>   0.0175 -0.2347 -0.2921  0.7611  0.0612  0.5934  0.6005 -0.2078 -0.6360
#>   0.1042 -0.0361 -0.3555 -0.4674  0.0108 -0.2792  0.4614 -0.0565 -0.5072
#> 
#> Columns 19 to 20  0.2270  0.4639
#>   0.5720  0.3737
#>  -0.2448 -0.2696
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0825 -0.5542 -0.2975  0.2610 -0.1212 -0.0370  0.0972 -0.0412  0.2512
#>  -0.5870 -0.6161 -0.0842  0.2615 -0.6310  0.0077  0.2649 -0.0330  0.4338
#>  -0.5503 -0.7839 -0.0151 -0.4665 -0.5299 -0.1716  0.4631 -0.7473  0.2117
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2224 -0.5516  0.0519 -0.1801 -0.0878  0.5621 -0.6415  0.2389 -0.4417
#>  -0.5234  0.2995  0.6096  0.5651 -0.3481  0.4313 -0.3961  0.2184  0.3754
#>  -0.3192 -0.5860 -0.2296 -0.7435  0.1737  0.7721 -0.0639  0.4565 -0.2176
#> 
#> Columns 10 to 18 -0.4141  0.6577 -0.2424  0.7264  0.8613 -0.4470 -0.4175 -0.6373  0.1524
#>  -0.6057 -0.0267  0.4212 -0.0206  0.4066  0.0454  0.2751 -0.1400 -0.2020
#>   0.6543  0.0945 -0.7077  0.7888  0.3318 -0.5973 -0.1967 -0.1969  0.0874
#> 
#> Columns 19 to 20  0.2777  0.1058
#>   0.4794 -0.3985
#>  -0.2705  0.1192
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5623 -0.6460  0.1007 -0.0147 -0.4487 -0.0862  0.2449 -0.1180  0.5182
#>  -0.0074 -0.2238  0.0734  0.1861  0.0160  0.5111 -0.0207 -0.0584  0.3952
#>  -0.3488 -0.6380  0.4749 -0.1135 -0.4764 -0.3101  0.5709 -0.5655  0.4981
#> 
#> Columns 10 to 18 -0.0279 -0.2377 -0.0953 -0.3637 -0.5493 -0.0875  0.5350 -0.2630 -0.1825
#>  -0.3926 -0.1192 -0.3301  0.2484  0.3683  0.3575  0.4972 -0.4377 -0.7074
#>  -0.2610 -0.3758  0.0172 -0.5896 -0.0341 -0.4729  0.7650 -0.4931  0.0901
#> 
#> Columns 19 to 20  0.2299  0.1211
#>   0.2918  0.0496
#>   0.2021 -0.3029
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>