RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8176 -0.2080  0.1550  0.3217 -0.5577  0.6754  0.5584  0.2088  0.3334
#>  -0.1654  0.8421 -0.5375 -0.8429  0.0760 -0.8492 -0.3294 -0.2669  0.6810
#>  -0.7994  0.0613 -0.1364  0.3138  0.2089  0.3351  0.9476 -0.2889  0.7460
#> 
#> Columns 10 to 18 -0.1821  0.3510  0.3596 -0.0680  0.6163 -0.6173 -0.8283  0.7357 -0.3119
#>   0.4461 -0.2863 -0.1438 -0.7569 -0.1602  0.5822 -0.5176  0.0817 -0.1407
#>  -0.5201  0.2201  0.0050  0.8247  0.0444  0.1655  0.6689  0.1304 -0.0666
#> 
#> Columns 19 to 20  0.5965 -0.4430
#>  -0.8401 -0.0602
#>  -0.0355  0.8395
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4389  0.7025 -0.3847  0.4327  0.0605  0.1878  0.7294 -0.3870  0.0345
#>   0.5021 -0.3972  0.5365  0.1394 -0.4186  0.1763 -0.0805 -0.3474  0.1427
#>   0.2099  0.4545 -0.2422  0.2103  0.2915 -0.2791  0.5910  0.2695  0.5161
#> 
#> Columns 10 to 18 -0.6344 -0.0502  0.4825  0.3660  0.6386  0.2161  0.2330 -0.1982 -0.0325
#>  -0.0579  0.5559 -0.0963  0.2385 -0.0688 -0.0790 -0.3344  0.5224  0.2544
#>  -0.2839  0.7843 -0.1405 -0.4523  0.3283  0.6346 -0.1978  0.0298 -0.5383
#> 
#> Columns 19 to 20  0.4533 -0.2140
#>  -0.1638  0.6300
#>  -0.0724 -0.7909
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2536  0.3778 -0.0202  0.3619 -0.0641 -0.2784  0.5117 -0.1886 -0.0253
#>   0.3319  0.6024 -0.1767 -0.0903 -0.0374 -0.0233  0.3382 -0.2768  0.3274
#>  -0.2497  0.4616 -0.4896  0.1366  0.2541 -0.0025  0.8496 -0.5529  0.0600
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1962  0.5947 -0.7373  0.4395 -0.3794  0.0529 -0.7040  0.5120 -0.0023
#>  -0.7554  0.0829 -0.6338 -0.3989 -0.3351  0.7920 -0.3916  0.3448 -0.5550
#>  -0.3503  0.1635  0.2378 -0.8188 -0.2511  0.2768  0.3269  0.3288 -0.4748
#> 
#> Columns 10 to 18  0.0164 -0.6792  0.3601  0.3431  0.1751  0.4068 -0.4731  0.1491  0.2273
#>  -0.0059 -0.4521  0.5140 -0.0763 -0.3593 -0.2464 -0.2818  0.6261 -0.5001
#>   0.0204 -0.4683  0.2145  0.4951 -0.8530 -0.8504  0.3494 -0.4159 -0.6184
#> 
#> Columns 19 to 20  0.7826 -0.1534
#>   0.2846 -0.4700
#>  -0.7246  0.2718
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5930  0.2535  0.0248  0.4734  0.1723  0.6610  0.5518 -0.2736  0.1363
#>  -0.0641  0.2359 -0.1342  0.2373  0.1015  0.1721  0.4457 -0.1056 -0.1165
#>  -0.2720  0.1481  0.1336  0.3764 -0.2116 -0.2375  0.0265 -0.1195 -0.1169
#> 
#> Columns 10 to 18  0.0647  0.4704 -0.2238 -0.3082  0.0417  0.0173 -0.3111  0.3728 -0.2694
#>   0.0408  0.6573 -0.0721  0.5522  0.3967  0.3689  0.0520 -0.3336 -0.3019
#>  -0.2773  0.7696  0.0660  0.6949  0.4426  0.5905 -0.1032  0.1671 -0.3047
#> 
#> Columns 19 to 20  0.1699  0.0006
#>  -0.0165 -0.1813
#>  -0.1729 -0.2758
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>