RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3455 -0.6115  0.3752  0.8479 -0.0229  0.2011 -0.4912 -0.2632 -0.3113
#>  -0.8881 -0.8756 -0.2235 -0.6155 -0.3502 -0.6674 -0.6739  0.3779 -0.2332
#>  -0.2338  0.3695  0.2457  0.6836  0.6528  0.0199 -0.1737 -0.4225  0.1610
#> 
#> Columns 10 to 18  0.4904  0.6605 -0.3497  0.1039  0.0297 -0.3952  0.1219 -0.4222 -0.7853
#>   0.2803  0.7040 -0.6761 -0.0253  0.0553  0.0349 -0.7349 -0.4704  0.7503
#>  -0.6424  0.7009 -0.6678  0.6027 -0.1511 -0.1302  0.6885 -0.3132 -0.2450
#> 
#> Columns 19 to 20 -0.6700  0.1076
#>   0.2005  0.1385
#>  -0.3513  0.2436
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2328  0.6500  0.2758  0.2880 -0.4035  0.2782 -0.2909  0.1968 -0.4294
#>   0.3933 -0.0625  0.4176 -0.0535 -0.5285 -0.1498 -0.3945  0.2365 -0.5049
#>  -0.4212  0.3827  0.0813  0.4371 -0.1837 -0.1602  0.2577 -0.1977  0.5890
#> 
#> Columns 10 to 18  0.6666  0.0562  0.5789 -0.6739  0.4382  0.0887 -0.1660 -0.2092  0.5947
#>   0.7622  0.0384  0.0234 -0.4952 -0.2791  0.3775 -0.1203  0.2745  0.4071
#>   0.5762 -0.2349  0.3027 -0.3915 -0.1286 -0.2273 -0.3783 -0.5498  0.5599
#> 
#> Columns 19 to 20  0.0943  0.1801
#>  -0.2966  0.4479
#>  -0.2382  0.2980
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3084  0.0013  0.5375  0.4387 -0.5637  0.1078  0.0205 -0.0630 -0.6262
#>  -0.5398  0.0720  0.0864  0.6965 -0.5524 -0.5602 -0.6037  0.3000 -0.6769
#>  -0.1447  0.2453  0.5194  0.4750 -0.0277 -0.0870 -0.2078  0.1260 -0.2074
#> 
#> Columns 10 to 18  0.7824 -0.2956 -0.4850 -0.3416  0.2657  0.3257 -0.0163  0.6600  0.3042
#>   0.5465 -0.3132 -0.5046 -0.1545  0.2353  0.1556 -0.4069  0.0507  0.7392
#>   0.7162  0.1121 -0.3212 -0.6840  0.2461  0.0119 -0.2956  0.0415  0.2177
#> 
#> Columns 19 to 20  0.4507  0.1345
#>  -0.2182 -0.3211
#>   0.2359  0.3290
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.3124  0.3204  0.2010  0.0470 -0.4861  0.4584 -0.7452 -0.2830  0.0377
#>   0.4990  0.0406  0.4619  0.4279 -0.6609  0.2358 -0.6796 -0.2951 -0.2605
#>   0.3894  0.3674  0.5024  0.2161 -0.4444  0.1494 -0.4602  0.0881 -0.4069
#> 
#> Columns 10 to 18  0.4767 -0.1890 -0.1004 -0.2266 -0.0458  0.2556  0.0113  0.5125  0.2069
#>   0.8549  0.3782 -0.2108 -0.3411  0.3528  0.1697 -0.5887  0.5735  0.5821
#>   0.8501  0.0289  0.0468 -0.2406  0.2449  0.2020 -0.2881  0.3377  0.2973
#> 
#> Columns 19 to 20 -0.1360 -0.0091
#>   0.4325  0.0818
#>   0.2184 -0.0880
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.3525 -0.0350 -0.0128  0.2615 -0.4431 -0.2768 -0.5025 -0.1179 -0.1635
#>   0.0209  0.0565  0.3142 -0.1513 -0.4111 -0.0497 -0.4938 -0.1534 -0.4579
#>   0.2007  0.0635  0.2045  0.0047 -0.5082 -0.4971 -0.3809 -0.1023 -0.3705
#> 
#> Columns 10 to 18  0.4850  0.4479 -0.3893 -0.4024  0.1382  0.1353 -0.4529  0.4971  0.1222
#>   0.7709  0.2078 -0.4280 -0.2760  0.1367  0.4828 -0.3777  0.6782 -0.0269
#>   0.6966 -0.0412 -0.2771 -0.0846  0.0788  0.6065 -0.3373  0.3354  0.1460
#> 
#> Columns 19 to 20  0.1739  0.0067
#>   0.4602  0.2058
#>  -0.1607  0.0317
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.3804 -0.7084  0.2395  0.4781 -0.0704  0.1894  0.2909 -0.4522 -0.2228
#>   0.1985 -0.4906 -0.2177 -0.2144  0.0380  0.3914  0.7055 -0.1778 -0.1559
#>   0.6165 -0.4166  0.2220  0.8368 -0.6957  0.7058 -0.3287  0.1491 -0.4744
#> 
#> Columns 10 to 18  0.5823  0.0828  0.4836  0.0108 -0.4014  0.1175 -0.7107 -0.4473  0.4349
#>  -0.1633 -0.4427 -0.1314 -0.5510 -0.5183  0.0046  0.2972  0.2100 -0.3957
#>  -0.1946  0.2511 -0.1199  0.0790  0.3319  0.6452 -0.0745  0.4645 -0.3895
#> 
#> Columns 19 to 20  0.5802 -0.1733
#>   0.7743  0.6107
#>   0.3727  0.4538
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3525 -0.0350 -0.0128  0.2615 -0.4431 -0.2768 -0.5025 -0.1179 -0.1635
#>   0.0209  0.0565  0.3142 -0.1513 -0.4111 -0.0497 -0.4938 -0.1534 -0.4579
#>   0.2007  0.0635  0.2045  0.0047 -0.5082 -0.4971 -0.3809 -0.1023 -0.3705
#> 
#> Columns 10 to 18  0.4850  0.4479 -0.3893 -0.4024  0.1382  0.1353 -0.4529  0.4971  0.1222
#>   0.7709  0.2078 -0.4280 -0.2760  0.1367  0.4828 -0.3777  0.6782 -0.0269
#>   0.6966 -0.0412 -0.2771 -0.0846  0.0788  0.6065 -0.3373  0.3354  0.1460
#> 
#> Columns 19 to 20  0.1739  0.0067
#>   0.4602  0.2058
#>  -0.1607  0.0317
#> [ CPUFloatType{2,3,20} ]
#>