RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0994  0.4144 -0.3580 -0.4605 -0.6749  0.3597  0.7566 -0.5026 -0.8376
#>  -0.7733  0.6336 -0.7516 -0.7092  0.3675  0.3143 -0.4477  0.5186  0.0699
#>   0.1818  0.0641  0.0027  0.7269 -0.0741  0.1835  0.5335 -0.6017 -0.0888
#> 
#> Columns 10 to 18  0.4095 -0.8100  0.0778  0.6165  0.6972  0.5188  0.2673 -0.5341  0.7284
#>  -0.4796 -0.3125  0.3484  0.8465  0.8985  0.8015  0.8512 -0.8764 -0.2456
#>  -0.4651 -0.0062  0.8959  0.4727  0.7160 -0.5748 -0.1038  0.3252  0.4123
#> 
#> Columns 19 to 20 -0.6999  0.1830
#>   0.9410  0.1862
#>   0.5286 -0.0447
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0777  0.2061  0.2344  0.1022 -0.5210  0.3737  0.4930 -0.1024 -0.8698
#>  -0.0281  0.5823  0.2768  0.7320 -0.4831  0.6948  0.0642 -0.4889 -0.3987
#>   0.0094  0.1708 -0.5433 -0.3517 -0.5666  0.3950  0.4289 -0.2310 -0.5991
#> 
#> Columns 10 to 18 -0.1905 -0.4289 -0.4169 -0.3078  0.2066 -0.0901 -0.3263  0.1456  0.5289
#>  -0.3389 -0.0916  0.4135  0.5724 -0.1754  0.0451  0.4942 -0.2002  0.2053
#>   0.3999 -0.5694 -0.1274  0.5828  0.1004  0.4254 -0.2087  0.1853  0.6051
#> 
#> Columns 19 to 20 -0.2609  0.4509
#>  -0.2894  0.4325
#>   0.4968  0.5412
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0356  0.2837  0.1400 -0.2775  0.3114  0.0399  0.2774  0.0941 -0.6028
#>   0.4325  0.1603 -0.0804  0.0710 -0.5428  0.5218  0.4581 -0.1451 -0.7041
#>   0.0591  0.3565  0.0239 -0.0157 -0.4970  0.6294  0.3713  0.2351 -0.8371
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8898 -0.2825 -0.3492  0.3361 -0.4917 -0.1919 -0.3720 -0.1317 -0.6524
#>   0.3026  0.0251  0.3544 -0.5107  0.1502  0.5015 -0.7074 -0.6699  0.0622
#>   0.3259 -0.0569 -0.1879 -0.0991 -0.1017  0.0624 -0.4704 -0.5758  0.3059
#> 
#> Columns 10 to 18 -0.0216 -0.5965 -0.1429 -0.6029 -0.6350 -0.7041  0.3313 -0.8516  0.6084
#>   0.2855 -0.5824  0.2255 -0.5455  0.0198 -0.4050  0.5053  0.4316 -0.2782
#>   0.1792 -0.5947 -0.2718 -0.3265  0.2203 -0.3721  0.2280  0.1196 -0.1193
#> 
#> Columns 19 to 20 -0.5039 -0.4734
#>   0.7071  0.6318
#>   0.1548 -0.0208
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2899  0.4952 -0.0389 -0.2426 -0.3700  0.5177  0.1602 -0.5852 -0.7015
#>   0.0421  0.1165 -0.0048 -0.3315 -0.2821  0.4604  0.3584  0.3825 -0.6948
#>  -0.0268  0.0354 -0.2024 -0.3135 -0.3463  0.2854  0.3859  0.2435 -0.8133
#> 
#> Columns 10 to 18  0.6150 -0.3691  0.3036  0.1266  0.3081 -0.1260 -0.3015  0.2996  0.1531
#>  -0.1905  0.0324 -0.1180 -0.0248  0.6804  0.6620  0.0472  0.1470  0.2072
#>  -0.1160  0.0264 -0.2169 -0.1772  0.5677  0.3840 -0.1818  0.1330 -0.1110
#> 
#> Columns 19 to 20  0.3290 -0.3072
#>  -0.1773  0.5512
#>  -0.2832  0.4379
#> [ CPUFloatType{2,3,20} ]
#>