RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1277 -0.5642  0.1040  0.5459 -0.0973  0.2306  0.3166 -0.0407  0.0103
#>   0.8045 -0.1612  0.7554  0.2098 -0.3275  0.5303  0.2383  0.0467 -0.3022
#>   0.8835  0.6207  0.5708  0.2197 -0.0682  0.8044  0.1704  0.6426 -0.0606
#> 
#> Columns 10 to 18  0.6557  0.2823 -0.0651 -0.2500 -0.6609 -0.2540 -0.1986  0.4836 -0.4713
#>  -0.5587 -0.5533 -0.7487  0.2071 -0.3875 -0.4201 -0.4250 -0.7046 -0.2999
#>  -0.8966 -0.6656 -0.6693  0.2002  0.1395 -0.5640 -0.0350 -0.1363 -0.0551
#> 
#> Columns 19 to 20  0.3400 -0.3517
#>   0.6771  0.8278
#>   0.6631  0.4240
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.7722  0.6172  0.3544 -0.4065 -0.1735 -0.2564  0.7157  0.5778  0.0910
#>   0.6894 -0.0914  0.6193  0.2747 -0.0620  0.1955  0.1442  0.0530 -0.5349
#>   0.6272  0.1723  0.3141  0.2636 -0.1777  0.6497  0.4934  0.5679 -0.3586
#> 
#> Columns 10 to 18 -0.0816  0.2730 -0.2273 -0.3286  0.1742 -0.1427  0.4246 -0.1925 -0.5615
#>  -0.3908 -0.5215 -0.4711  0.2339  0.3517 -0.3976 -0.4595 -0.4254  0.2551
#>  -0.0362 -0.1664 -0.4059  0.3255  0.1987 -0.0105  0.1473 -0.5403 -0.3003
#> 
#> Columns 19 to 20  0.1038  0.4658
#>   0.7274 -0.3935
#>   0.7002 -0.3326
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2296  0.4053  0.1806  0.1841 -0.4059  0.6547 -0.4478  0.9206  0.3603
#>   0.4475 -0.0107  0.6653  0.6046 -0.0718  0.1746  0.6779 -0.2259 -0.2031
#>   0.1081  0.0827  0.4921  0.3188 -0.1799  0.5067  0.1077  0.2866 -0.0103
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2807  0.4340 -0.5225  0.2220  0.4752 -0.1983 -0.0908  0.5929 -0.5395
#>  -0.3833  0.1163  0.1372  0.1390  0.0480 -0.2442 -0.4940  0.1568 -0.1443
#>  -0.2708 -0.0486 -0.1324  0.6026 -0.1902  0.1054 -0.4833  0.1517 -0.4110
#> 
#> Columns 10 to 18  0.2218  0.3416 -0.3353  0.1415 -0.1541  0.0138 -0.2787  0.3823 -0.4453
#>  -0.1860 -0.1669  0.4483  0.2329  0.3669 -0.1532 -0.1536  0.1832 -0.3362
#>  -0.2498 -0.0952  0.4237 -0.1080  0.5808 -0.3709 -0.0391  0.6045  0.3860
#> 
#> Columns 19 to 20  0.6808 -0.3841
#>   0.1913 -0.3614
#>  -0.2195 -0.4963
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5205  0.5114  0.5048  0.6181 -0.3602  0.2036  0.2321  0.1648  0.2308
#>   0.3481  0.4281  0.3447  0.2343 -0.1762  0.2308  0.2195  0.5329 -0.1949
#>   0.3287  0.4195  0.4225  0.3028 -0.1271  0.2772  0.2609  0.4691 -0.0542
#> 
#> Columns 10 to 18 -0.6167 -0.4133  0.1260  0.0522  0.3316 -0.4292 -0.2366 -0.0580  0.2733
#>  -0.4954 -0.2253  0.0556 -0.3820 -0.0861  0.0272  0.0826 -0.0727 -0.2413
#>  -0.6078 -0.4407  0.0787 -0.6307  0.1484 -0.1307  0.1025  0.2771 -0.2379
#> 
#> Columns 19 to 20  0.6561  0.1063
#>   0.6642  0.0511
#>   0.7642 -0.0219
#> [ CPUFloatType{2,3,20} ]
#>