RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0412  0.7955 -0.8508 -0.1714  0.1932  0.7105  0.3664 -0.2765 -0.4922
#>   0.3342  0.6815  0.7798  0.8792  0.8847 -0.3766 -0.4441 -0.7595 -0.3607
#>   0.7635  0.4127 -0.3795  0.1237 -0.5200  0.2981 -0.4667 -0.7109 -0.4041
#> 
#> Columns 10 to 18 -0.4356  0.6928 -0.0487 -0.0788 -0.5001 -0.4978 -0.7463  0.3177 -0.5092
#>  -0.0095  0.6598  0.4257 -0.5679  0.1370 -0.6765  0.6427  0.9008 -0.6627
#>  -0.2097  0.4275 -0.6638 -0.8582 -0.3957 -0.0640  0.4437  0.1132  0.4013
#> 
#> Columns 19 to 20  0.5815  0.1529
#>   0.7910  0.1839
#>   0.5694 -0.7985
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2545  0.5656  0.1459 -0.6727 -0.2811  0.1778 -0.0339  0.2887  0.0374
#>   0.2527 -0.4243 -0.0482 -0.3564  0.0692  0.5853  0.1096 -0.2105 -0.6656
#>   0.6500  0.2835 -0.0816 -0.4672 -0.1965  0.1353  0.0351  0.0861 -0.3426
#> 
#> Columns 10 to 18  0.1647  0.2648  0.0880  0.2465 -0.4406  0.1888  0.2847  0.1237 -0.4180
#>  -0.1129 -0.6412  0.1753 -0.3452 -0.3593 -0.2278  0.3000  0.1333 -0.7012
#>   0.3611 -0.1119  0.0488  0.1629 -0.1142 -0.1779 -0.0981 -0.1403 -0.3859
#> 
#> Columns 19 to 20  0.1686 -0.0444
#>   0.0654 -0.5866
#>  -0.0570  0.0170
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1943  0.0341  0.1422 -0.1902 -0.2482 -0.2741 -0.2865 -0.2975 -0.2194
#>   0.4561  0.2233  0.2453 -0.6036  0.0479  0.1586 -0.5812  0.1919 -0.8099
#>   0.1347  0.5468 -0.1993 -0.3365 -0.2962 -0.0527 -0.0834  0.2091  0.0311
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2002 -0.7494 -0.8232  0.8068  0.5662 -0.1128  0.7007  0.0156 -0.2080
#>   0.4690  0.0370 -0.5223  0.3900  0.5931  0.2757  0.1634 -0.0813 -0.0463
#>  -0.5883  0.1193  0.3105  0.3125 -0.2099  0.2650  0.1951 -0.3684 -0.3061
#> 
#> Columns 10 to 18  0.7253  0.4816  0.4689 -0.4052  0.4304  0.5139 -0.1743 -0.6644 -0.4326
#>   0.4058 -0.3235  0.6725 -0.5468  0.2389  0.1108  0.3935  0.0487  0.4435
#>  -0.3123  0.1452 -0.4197  0.3051  0.1635  0.3496 -0.1053  0.0104 -0.5074
#> 
#> Columns 19 to 20  0.6052  0.7008
#>  -0.1880  0.0793
#>  -0.4150  0.4322
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0333 -0.3727  0.4100 -0.1639 -0.3905  0.0231 -0.5325 -0.2354 -0.5153
#>  -0.1606  0.3193 -0.0072 -0.3878 -0.0365  0.0458 -0.3710  0.0328 -0.3572
#>   0.1159  0.1154  0.3326 -0.2690 -0.4240  0.1359  0.0301 -0.1415 -0.3776
#> 
#> Columns 10 to 18 -0.6601 -0.4359  0.1024 -0.1635 -0.2141 -0.2491  0.3614  0.3863 -0.4047
#>  -0.2081 -0.1617 -0.0907  0.0360 -0.2833 -0.2424  0.1366 -0.1613 -0.1718
#>  -0.2813  0.4066  0.3102  0.0030 -0.1122 -0.1571  0.4486  0.4832 -0.4445
#> 
#> Columns 19 to 20 -0.2901 -0.1386
#>  -0.0975 -0.0179
#>   0.0976 -0.0459
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>