RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4006 -0.4559 -0.2828  0.4823 -0.4096  0.8162 -0.8819  0.6390 -0.1299
#>  -0.3964 -0.5470 -0.3560 -0.6995 -0.3496 -0.5758  0.5168 -0.1536 -0.3894
#>   0.3070 -0.3892 -0.6982 -0.7978  0.5688  0.2204  0.0493  0.4187 -0.4540
#> 
#> Columns 10 to 18 -0.2067  0.5202  0.1810  0.0860 -0.3225  0.7887 -0.7162 -0.4859 -0.2192
#>   0.4335 -0.3468 -0.0180 -0.0988 -0.7166 -0.2470  0.0036 -0.6401 -0.0538
#>   0.6656 -0.4296 -0.4430  0.7455  0.3551  0.2738 -0.5250 -0.5848  0.1567
#> 
#> Columns 19 to 20  0.9321  0.2083
#>   0.1297 -0.2031
#>   0.7468  0.0881
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4013 -0.3446  0.0715 -0.1860 -0.6468 -0.0322 -0.5650  0.3095 -0.2085
#>   0.0456  0.2460 -0.5555 -0.5948 -0.1665  0.1432 -0.0741 -0.0641  0.0402
#>   0.3469 -0.0277  0.3889  0.2756  0.0486  0.2718  0.2766  0.2038 -0.1444
#> 
#> Columns 10 to 18  0.5765 -0.6374 -0.3098 -0.1849 -0.6457  0.0447  0.1024 -0.0244  0.3949
#>   0.4404 -0.3519  0.1144  0.7598 -0.2985  0.3514  0.5698  0.1159  0.3295
#>   0.7179  0.2056  0.2549  0.6474 -0.3245 -0.0936  0.5213 -0.2848  0.5734
#> 
#> Columns 19 to 20  0.0243 -0.2202
#>   0.3563  0.0273
#>   0.0415 -0.5814
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2839  0.3890 -0.1817 -0.2274 -0.2558 -0.3626  0.0130 -0.1375 -0.5204
#>  -0.5150  0.3125 -0.3378 -0.3163 -0.3577  0.1435 -0.2548 -0.2355  0.1215
#>  -0.5624  0.1014 -0.4856 -0.4532 -0.6443 -0.5090  0.0942 -0.3635  0.0113
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5182  0.5760 -0.3304 -0.3722  0.6942 -0.1412 -0.5713 -0.3724 -0.6626
#>  -0.1932 -0.1567 -0.0492  0.0408 -0.0083 -0.2653 -0.3046 -0.3039  0.1829
#>  -0.2167 -0.5888 -0.4643 -0.1014  0.0989 -0.6160  0.2395 -0.5489  0.2615
#> 
#> Columns 10 to 18  0.5269 -0.1208  0.5661  0.3511 -0.1745 -0.0047  0.1167 -0.1798 -0.8135
#>   0.0931 -0.3053  0.3510  0.4099  0.0288 -0.0698 -0.0937  0.2611 -0.0734
#>  -0.2671 -0.3416 -0.1087  0.3188  0.2515 -0.4061 -0.4588 -0.0070 -0.3885
#> 
#> Columns 19 to 20  0.0525  0.5704
#>   0.3866  0.2448
#>   0.0088  0.1830
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1931 -0.4280  0.0217 -0.3870 -0.2854 -0.3255  0.3311 -0.0999 -0.1756
#>   0.0303 -0.0961  0.0349 -0.6282 -0.2554 -0.0502  0.0533 -0.1685  0.0728
#>   0.2303 -0.1550  0.1862 -0.1982 -0.4863  0.1616  0.0866  0.0130 -0.2359
#> 
#> Columns 10 to 18  0.5295 -0.4440  0.3871  0.1712 -0.4061  0.0671  0.4723 -0.3326  0.3908
#>   0.5342 -0.0054  0.4496  0.3521 -0.4713  0.2606  0.2929 -0.0294 -0.0711
#>   0.2311  0.1264  0.3325  0.1193 -0.2776  0.0416  0.0551 -0.3250  0.1159
#> 
#> Columns 19 to 20  0.0328 -0.5550
#>  -0.3647 -0.3331
#>   0.5150 -0.3433
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>