RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0566 -0.8940 -0.5649 -0.2074 -0.8827  0.3344  0.3742 -0.5628 -0.6227
#>   0.3170  0.5213  0.8685  0.3388  0.4106 -0.8258 -0.7340 -0.1551 -0.2310
#>  -0.5151  0.7593  0.2706  0.5380 -0.8304 -0.3196 -0.7392 -0.7942 -0.4340
#> 
#> Columns 10 to 18 -0.0768 -0.8614 -0.3869 -0.1421  0.8766  0.7693  0.3940 -0.6908 -0.2104
#>  -0.5347  0.4443 -0.0475  0.7016  0.4398  0.6105 -0.8665 -0.1528  0.7877
#>   0.0132  0.8436  0.3436 -0.7541 -0.5560 -0.7877  0.9405  0.6296  0.4725
#> 
#> Columns 19 to 20  0.4799  0.4363
#>   0.6029  0.7759
#>   0.4763 -0.4294
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7575 -0.6242  0.0091 -0.0393 -0.1576 -0.1201  0.3629  0.3416 -0.1409
#>  -0.3245 -0.0172  0.3268 -0.2709  0.0445 -0.3295  0.4365  0.0315 -0.1127
#>  -0.3778  0.2217  0.5137  0.4879 -0.5498  0.3872  0.5271 -0.2364  0.2009
#> 
#> Columns 10 to 18  0.3029  0.0054  0.5364  0.2955  0.5063  0.3296 -0.3180 -0.4528 -0.2875
#>  -0.1221 -0.4885  0.1924  0.7596  0.1063  0.1021 -0.7586 -0.4231 -0.6219
#>   0.2743 -0.3174 -0.0892 -0.5647  0.1918  0.2571  0.4913 -0.3744  0.0729
#> 
#> Columns 19 to 20 -0.1262  0.4809
#>   0.0711  0.4187
#>  -0.2297  0.5465
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5876 -0.3042  0.2968 -0.1929 -0.4609 -0.2478  0.4673  0.2602 -0.3877
#>  -0.7046 -0.3799  0.5102  0.0025 -0.2165 -0.1481  0.6654  0.0989 -0.5874
#>  -0.5275  0.1407  0.6196  0.1154 -0.1408 -0.0067 -0.0389  0.1840 -0.2530
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4271  0.1387  0.6975 -0.5037  0.1980 -0.0217 -0.7893  0.4909 -0.5096
#>  -0.0585  0.2394 -0.5670  0.0107  0.3118 -0.6529  0.6273 -0.3652 -0.0591
#>   0.5125  0.2879  0.3611 -0.2538  0.1506 -0.2894 -0.7648  0.1979 -0.3759
#> 
#> Columns 10 to 18 -0.4048  0.8481  0.5387  0.4253  0.8798 -0.0490  0.5808 -0.1776 -0.4072
#>  -0.7170 -0.2486 -0.2821  0.1792  0.0952  0.2488 -0.7468 -0.1681  0.6446
#>  -0.5981  0.6384  0.6436  0.0565  0.7202  0.5413  0.6794 -0.4566 -0.2889
#> 
#> Columns 19 to 20  0.3248 -0.1008
#>   0.3628  0.0164
#>   0.6481 -0.0387
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5052 -0.1664  0.7433 -0.0513 -0.4325  0.0999  0.6190 -0.1945 -0.7022
#>  -0.1707  0.0210  0.2757  0.0222 -0.3961 -0.6000  0.3027 -0.0698 -0.0471
#>  -0.5701 -0.0678  0.5143 -0.0411 -0.5979 -0.0568  0.3788 -0.1323 -0.7424
#> 
#> Columns 10 to 18  0.6103  0.0527 -0.0261  0.3592  0.2522 -0.3503  0.0308 -0.4668  0.5742
#>  -0.1953  0.0385 -0.3361  0.3076  0.3162  0.1583  0.1661 -0.2951  0.2099
#>   0.2975 -0.0456  0.1784  0.2314  0.2517 -0.5026  0.2673 -0.0597  0.5241
#> 
#> Columns 19 to 20 -0.1024  0.2064
#>  -0.0296 -0.0352
#>  -0.2150 -0.2227
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>