RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5118  0.2358  0.4813 -0.5991  0.6262  0.3531 -0.2711 -0.5909  0.5730
#>  -0.4613  0.0590 -0.4130 -0.7243  0.2952 -0.0046  0.6457 -0.7217  0.2838
#>   0.8134  0.5794 -0.2165  0.0090  0.1075 -0.3150  0.0193  0.3334 -0.3787
#> 
#> Columns 10 to 18 -0.3564 -0.1384  0.3440 -0.3865 -0.0848  0.0828  0.5730  0.2156 -0.6443
#>   0.4531 -0.6847 -0.6809  0.6491 -0.6365 -0.0826  0.2330 -0.0505  0.2496
#>   0.0539 -0.2474  0.8760  0.1973 -0.6880 -0.1746  0.4974 -0.2446 -0.7762
#> 
#> Columns 19 to 20  0.5397  0.1138
#>  -0.7827 -0.0444
#>   0.3472  0.0219
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2461  0.1303 -0.1952  0.0657 -0.4122 -0.2445 -0.2013 -0.0357  0.4610
#>   0.6919 -0.5645 -0.6741  0.4617  0.5453  0.1447 -0.0271 -0.0878  0.1515
#>   0.1433 -0.1353 -0.6038 -0.5343  0.5465 -0.3104 -0.0774 -0.2316  0.1356
#> 
#> Columns 10 to 18  0.1824 -0.1357  0.3264  0.1907 -0.4983  0.4065 -0.2763  0.6145 -0.5254
#>   0.2565  0.1561  0.6550 -0.0388 -0.3484 -0.0848  0.7091 -0.4076 -0.5470
#>   0.1569 -0.1360  0.5098  0.1961  0.1307  0.2872  0.1467  0.1087 -0.6277
#> 
#> Columns 19 to 20 -0.3891 -0.2259
#>   0.3651  0.3672
#>  -0.3558  0.0353
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1142 -0.0513 -0.6209  0.3687  0.1690 -0.4254  0.0394  0.1720  0.2147
#>   0.2683  0.0454 -0.3518  0.0465 -0.3805 -0.6003 -0.2444 -0.0421 -0.3697
#>   0.0342 -0.3964 -0.6174 -0.3428  0.5335 -0.0494  0.4152 -0.0833  0.6573
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0623 -0.0943  0.4055 -0.1206  0.3835 -0.1406  0.2279  0.1248  0.1816
#>   0.5715  0.1356 -0.2728 -0.4489  0.3801 -0.1221 -0.7487  0.2194  0.0672
#>   0.3045 -0.4965 -0.2328 -0.4245  0.2906  0.2351  0.0399 -0.0483  0.4105
#> 
#> Columns 10 to 18  0.0696  0.4166 -0.1567 -0.4187 -0.0918  0.1753  0.4667 -0.3646 -0.4364
#>   0.4830  0.2455  0.5477 -0.4181 -0.7747 -0.0786  0.5219 -0.2555 -0.4630
#>  -0.3312 -0.5481  0.1118 -0.2461  0.4734 -0.0901 -0.1861 -0.4275 -0.1337
#> 
#> Columns 19 to 20  0.3037 -0.3547
#>   0.2937 -0.0455
#>   0.6744  0.2709
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0957  0.2193 -0.6122 -0.0226 -0.0460 -0.2665  0.2748  0.2376  0.1992
#>  -0.2853  0.4006 -0.6847  0.3722 -0.3675 -0.1432 -0.0046  0.4175  0.0731
#>   0.3242 -0.0183 -0.3483 -0.1822 -0.1275 -0.4111  0.1301 -0.2022  0.4911
#> 
#> Columns 10 to 18 -0.5349  0.2002  0.2754 -0.1151 -0.2578 -0.0971  0.3158  0.4135 -0.0386
#>   0.1801  0.0420  0.0552  0.2395  0.2535  0.4462  0.3865  0.7056  0.2732
#>  -0.4151  0.2318  0.7131  0.1881 -0.0719  0.1859 -0.0470  0.0140 -0.1116
#> 
#> Columns 19 to 20 -0.4796 -0.1382
#>  -0.6767 -0.1484
#>  -0.0344 -0.0549
#> [ CPUFloatType{2,3,20} ]
#>