RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8268 -0.6424  0.8395  0.4035  0.6827  0.2934  0.3760 -0.9395 -0.2868
#>  -0.7229  0.7131 -0.0320 -0.0925  0.3020 -0.4912  0.7179 -0.3004 -0.4250
#>  -0.5144 -0.2622  0.1431 -0.3328 -0.0773 -0.0377 -0.2849 -0.5696  0.1975
#> 
#> Columns 10 to 18  0.0234  0.5717 -0.7503  0.1201 -0.8084 -0.4090  0.6438 -0.8414 -0.6389
#>  -0.3029  0.2373  0.6502 -0.3895 -0.4004 -0.6719  0.8491 -0.6921 -0.0197
#>  -0.9290 -0.3099  0.3215  0.6269  0.8402 -0.1680 -0.4709 -0.7139 -0.8660
#> 
#> Columns 19 to 20 -0.8591 -0.6642
#>  -0.4070 -0.4154
#>   0.3937 -0.5800
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5835 -0.1138  0.1234  0.3826 -0.1504  0.2959  0.0359 -0.5641 -0.5593
#>  -0.8761  0.2866 -0.0371 -0.3366 -0.4344  0.3095 -0.3848 -0.0726 -0.3163
#>  -0.0556  0.3474  0.2159  0.4382  0.4353 -0.0531  0.0646 -0.5437 -0.3489
#> 
#> Columns 10 to 18 -0.5520 -0.4638 -0.4052  0.4616  0.6871  0.3369  0.3090 -0.3592 -0.7114
#>  -0.0636 -0.1075 -0.2430  0.4042  0.8410  0.3510 -0.1956  0.1091 -0.2240
#>   0.3866  0.5678 -0.5395 -0.3281  0.2041  0.5134 -0.0819 -0.3576 -0.2626
#> 
#> Columns 19 to 20 -0.2079 -0.4625
#>   0.4996  0.1012
#>  -0.1980 -0.2432
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.6755  0.5117  0.6705  0.0475  0.2239 -0.2486  0.1055 -0.4112 -0.2887
#>  -0.2296  0.7533 -0.6126  0.7968 -0.1480  0.1751  0.3554  0.1610  0.0341
#>  -0.7815  0.4918  0.1217  0.4001 -0.4401  0.1114  0.2208 -0.2448  0.0628
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1935 -0.0541 -0.5164 -0.1992 -0.0990 -0.4358  0.0873  0.0316 -0.6528
#>   0.8542 -0.7549  0.4100 -0.0836 -0.2566 -0.5096  0.4870  0.6552 -0.3569
#>   0.1304  0.0222  0.6835 -0.1709 -0.8094 -0.1096  0.0998 -0.6036  0.7451
#> 
#> Columns 10 to 18 -0.0761  0.6510 -0.0526 -0.2310 -0.3010 -0.0271 -0.4167 -0.1819  0.0139
#>  -0.4901  0.2959 -0.7301  0.1807  0.0160  0.2278  0.3382 -0.0331 -0.0512
#>  -0.9559  0.0040  0.7571  0.8244  0.1759 -0.7105 -0.1855  0.5688  0.2774
#> 
#> Columns 19 to 20 -0.5303 -0.4948
#>   0.5113 -0.0127
#>  -0.1825 -0.4958
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0177  0.4954 -0.2438  0.4752  0.2120  0.3773 -0.0005  0.1559  0.2239
#>  -0.1191 -0.0515  0.2584  0.6687  0.1945  0.6265 -0.2601  0.1270 -0.0453
#>  -0.2185  0.5052 -0.3652  0.6631 -0.0253  0.2418  0.4396  0.2912  0.2178
#> 
#> Columns 10 to 18 -0.0614  0.0104 -0.0159 -0.2844  0.6342  0.7239 -0.0929 -0.1859 -0.1663
#>  -0.3471  0.0493 -0.2680 -0.2057  0.2764  0.3002 -0.1766 -0.5293 -0.3397
#>  -0.1837  0.3811  0.1696 -0.6990 -0.0635 -0.2476  0.3935 -0.4624 -0.2762
#> 
#> Columns 19 to 20  0.1261 -0.0778
#>  -0.0407 -0.0096
#>  -0.3815 -0.3848
#> [ CPUFloatType{2,3,20} ]
#>