RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3310 -0.1263  0.4329 -0.0096 -0.6357  0.2913 -0.6303 -0.2335 -0.1005
#>   0.2451  0.3557 -0.1705  0.2733  0.3957 -0.3583  0.5659  0.8833  0.5648
#>  -0.6244  0.4556  0.0419 -0.8747 -0.2414  0.2435 -0.1508 -0.3809  0.1626
#> 
#> Columns 10 to 18  0.2255 -0.4026  0.2447  0.7459 -0.3982  0.1568  0.8951 -0.3343 -0.3852
#>  -0.7887  0.7699  0.0592 -0.3487 -0.4436  0.7583  0.6931 -0.5197 -0.6294
#>  -0.0866  0.4388  0.4408 -0.2900 -0.4464 -0.3608 -0.1175 -0.1050 -0.9055
#> 
#> Columns 19 to 20 -0.3424  0.2127
#>  -0.2742 -0.5855
#>   0.7286  0.1991
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2178  0.0537 -0.5312  0.4363  0.5825 -0.3606  0.4168  0.2870 -0.1569
#>  -0.4739 -0.1888  0.0377  0.6657  0.4460 -0.2808 -0.3475  0.1170 -0.6515
#>  -0.0323  0.0187 -0.3543 -0.5599 -0.3442  0.1203 -0.0009  0.0566  0.0191
#> 
#> Columns 10 to 18 -0.0206 -0.0493  0.3172 -0.4120 -0.1848  0.4649 -0.3491  0.4495 -0.4819
#>   0.0492 -0.9516  0.5785  0.3860 -0.1520  0.6213  0.3271  0.0860  0.3445
#>  -0.2822 -0.2828  0.5233 -0.3490 -0.0172 -0.4745  0.2829 -0.2064 -0.4274
#> 
#> Columns 19 to 20  0.1934 -0.2146
#>   0.0645  0.0291
#>   0.3560 -0.2749
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0369 -0.4855 -0.0057  0.0869 -0.3115 -0.0758  0.1205  0.4004  0.3324
#>   0.0710 -0.0742 -0.2129  0.0499  0.0632 -0.3912  0.4745  0.2259 -0.0454
#>  -0.4047 -0.2703 -0.5989 -0.0671  0.1638  0.3768  0.3689  0.4969  0.2371
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6442 -0.3051  0.1668  0.3888 -0.1334  0.3165  0.5051 -0.5905 -0.2677
#>  -0.4726 -0.3522  0.2586 -0.0042  0.4656 -0.3339  0.2685 -0.3112 -0.1448
#>  -0.3926 -0.7406  0.7469  0.6821  0.0754 -0.3133  0.2577  0.4501  0.6077
#> 
#> Columns 10 to 18 -0.4478  0.5868  0.2568  0.3343  0.1981 -0.2585 -0.7962 -0.3402 -0.4611
#>  -0.3737 -0.3438 -0.4593  0.5086  0.3488 -0.1924 -0.1897 -0.5786  0.1440
#>  -0.4738  0.6722 -0.5097  0.7061 -0.2249 -0.5753  0.0744 -0.6426  0.1722
#> 
#> Columns 19 to 20 -0.0627  0.6386
#>  -0.6311 -0.0380
#>   0.2406  0.4444
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2843  0.2705  0.0091  0.3387 -0.0222  0.2557  0.0705  0.2600 -0.1851
#>  -0.4083  0.5720 -0.0521  0.1169  0.0081  0.0295 -0.1580 -0.0055 -0.0188
#>  -0.7569  0.0550 -0.4317 -0.0834  0.2553  0.2478 -0.0827  0.4252  0.1530
#> 
#> Columns 10 to 18  0.1454 -0.3225  0.3902 -0.2340 -0.2018  0.3929 -0.2507  0.0779 -0.3817
#>   0.5014 -0.2748  0.1772 -0.1662 -0.0961  0.3492 -0.3870 -0.1785 -0.4971
#>  -0.0288 -0.6062 -0.1329 -0.0181 -0.0361  0.2673 -0.0956 -0.1365 -0.3048
#> 
#> Columns 19 to 20  0.1151 -0.2553
#>  -0.0567 -0.1263
#>   0.2090 -0.2378
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>