RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1015  0.3281  0.9564 -0.0165 -0.2314  0.2294  0.1988 -0.3256  0.8518
#>  -0.3785 -0.8486 -0.6800 -0.3453  0.5511  0.6514  0.1076 -0.1099 -0.5006
#>  -0.4393  0.4984  0.7301  0.3379  0.2450 -0.4727  0.5116 -0.1278 -0.1654
#> 
#> Columns 10 to 18  0.5205 -0.6445 -0.2068 -0.4302 -0.1555  0.5649  0.4887  0.3446 -0.1295
#>   0.2086  0.2816 -0.3689  0.6790 -0.0280  0.7697  0.0124  0.6853  0.8161
#>   0.5013 -0.6696  0.4276 -0.4487 -0.3389 -0.6201 -0.4077  0.8230 -0.4513
#> 
#> Columns 19 to 20  0.0409  0.2695
#>  -0.6301  0.4548
#>   0.3323 -0.0879
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0934 -0.1863  0.4123  0.3687 -0.2112 -0.4719 -0.4601  0.0013 -0.4988
#>  -0.5461 -0.6646  0.3892 -0.4674 -0.6487  0.3500 -0.4971  0.0232  0.2211
#>  -0.1082 -0.2949 -0.1431 -0.1199 -0.4143 -0.3344 -0.1060 -0.1288 -0.5970
#> 
#> Columns 10 to 18  0.4908 -0.6052 -0.1774  0.0118 -0.0446 -0.2241  0.1347  0.5438  0.5511
#>   0.5254 -0.2300 -0.3445 -0.3455 -0.2638  0.6458  0.1247  0.1709 -0.5212
#>   0.4163 -0.2101  0.3454  0.4418 -0.0000  0.2763  0.4218  0.3798  0.7248
#> 
#> Columns 19 to 20  0.0726 -0.2191
#>   0.3651  0.1174
#>   0.2157 -0.1735
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1956 -0.3748 -0.2327 -0.1031  0.2465 -0.1489  0.4594  0.2254  0.1148
#>  -0.1598 -0.1312  0.2265 -0.1078 -0.3730 -0.1466 -0.6892 -0.3560  0.0783
#>  -0.2821 -0.3043  0.2020 -0.0652 -0.2742  0.0952  0.2643 -0.1508  0.3641
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3521 -0.0574  0.0419 -0.2697  0.1331  0.2379 -0.2765  0.2287 -0.1523
#>  -0.2628 -0.2635  0.5689  0.0849 -0.0955 -0.7152 -0.1105 -0.4692 -0.0602
#>  -0.5269 -0.2327  0.5442 -0.5808  0.3513 -0.3958  0.4356 -0.4826 -0.3654
#> 
#> Columns 10 to 18  0.0452 -0.2684  0.2058 -0.3926 -0.7107  0.1021  0.3045 -0.4775  0.3131
#>  -0.5835 -0.1821 -0.3209 -0.1098 -0.5747  0.1132  0.1740 -0.0034  0.2658
#>  -0.0782 -0.0423 -0.0473 -0.2826 -0.6682  0.1185  0.3941 -0.3835  0.2507
#> 
#> Columns 19 to 20  0.6451  0.2335
#>   0.3014  0.2649
#>   0.6455  0.0804
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1017 -0.0648  0.0684 -0.2409 -0.0458 -0.0452  0.1115  0.1556  0.5584
#>  -0.3022 -0.4231  0.1621  0.1599 -0.1779  0.5238 -0.0588  0.0744  0.0545
#>  -0.1205 -0.5928 -0.0678 -0.4400 -0.4030  0.6091  0.1040  0.1374  0.3034
#> 
#> Columns 10 to 18  0.6138 -0.1375 -0.2320  0.2038 -0.3154  0.1353  0.2325  0.1215  0.2934
#>   0.3040  0.1127  0.2567 -0.1093 -0.2747  0.1385 -0.1176  0.0808  0.3374
#>   0.6032  0.3097  0.0685 -0.0961 -0.4259  0.2828  0.2516  0.0635  0.3849
#> 
#> Columns 19 to 20  0.4613 -0.0315
#>   0.1265  0.1700
#>   0.4896  0.2162
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>