RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5660 -0.7451  0.7140  0.6869  0.3612  0.5140 -0.9308  0.5898 -0.1323
#>   0.0915  0.5832 -0.8869  0.8327 -0.8412 -0.8819  0.5237 -0.0166  0.7458
#>   0.5950  0.8890  0.5696  0.8377  0.6319  0.2206 -0.6729 -0.2332  0.1745
#> 
#> Columns 10 to 18 -0.6295  0.4654 -0.4166  0.5127  0.0822  0.6918 -0.9442 -0.3456  0.6632
#>  -0.9584 -0.6365 -0.3657 -0.5444 -0.9310 -0.9146 -0.0694  0.6133  0.8294
#>  -0.1398  0.4836  0.0450 -0.7995 -0.3831  0.5473  0.3702 -0.1667  0.0015
#> 
#> Columns 19 to 20 -0.5423 -0.9579
#>  -0.2509 -0.6381
#>   0.4493 -0.2545
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1880  0.4069 -0.5663  0.4680 -0.4502 -0.4675 -0.2380 -0.2402  0.2645
#>  -0.0666  0.3551  0.6336 -0.1605 -0.4251  0.1094  0.5445 -0.1796  0.0656
#>   0.1257  0.1625  0.0664 -0.2046  0.1292  0.1952 -0.5189  0.4001 -0.4956
#> 
#> Columns 10 to 18 -0.2026  0.2295  0.2906 -0.3130 -0.5008 -0.1623 -0.6140 -0.2704  0.1845
#>  -0.2370  0.0008  0.8241 -0.6163 -0.4512 -0.5519 -0.3260  0.2555  0.2282
#>  -0.1829 -0.1662 -0.0186  0.0663 -0.6741 -0.1543 -0.4438  0.0434 -0.0168
#> 
#> Columns 19 to 20 -0.1213 -0.8118
#>  -0.6294 -0.5145
#>  -0.2677 -0.5840
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2536  0.2531  0.0517  0.6366  0.3935  0.0010  0.1645 -0.1665  0.0347
#>  -0.1705  0.6110  0.1502  0.3980  0.1102  0.0361  0.1951  0.5564  0.0426
#>  -0.0243  0.5852  0.2211  0.3117 -0.4242 -0.0766 -0.0011 -0.0666 -0.1226
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1835 -0.6101  0.1305 -0.3123  0.5537 -0.5164 -0.1660  0.6470 -0.3619
#>   0.2905 -0.2211  0.0434  0.5638 -0.4978 -0.0427  0.0841  0.1367  0.2430
#>   0.3926 -0.0345  0.2629 -0.1963 -0.1072 -0.5869 -0.7921  0.3252 -0.6917
#> 
#> Columns 10 to 18  0.3624 -0.2366  0.5838 -0.1674 -0.3099  0.0059  0.2317 -0.7138  0.4487
#>  -0.3473 -0.6214 -0.0216 -0.1124  0.2011 -0.4235  0.8429  0.4983  0.1472
#>   0.2120 -0.5947  0.6250  0.4114 -0.4944 -0.0700  0.8638 -0.0926 -0.5207
#> 
#> Columns 19 to 20  0.0366  0.1320
#>   0.5713  0.2495
#>  -0.1438  0.8473
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2201  0.3936  0.2469  0.5637 -0.0971  0.3606  0.2498 -0.2145  0.0443
#>  -0.1649  0.5524 -0.3413  0.0646 -0.4692 -0.1659  0.1103  0.2852  0.3324
#>  -0.0676  0.7327 -0.2730  0.4983 -0.2015 -0.2871 -0.0455  0.3829  0.2214
#> 
#> Columns 10 to 18  0.2138  0.2931  0.2592 -0.4741  0.1236  0.1165 -0.6097 -0.1182  0.1678
#>  -0.3734  0.4558  0.0383 -0.3350 -0.4589 -0.0647 -0.3619  0.4230  0.4384
#>   0.3297 -0.0921  0.1219 -0.2196 -0.0512 -0.1073 -0.5107 -0.1620 -0.0656
#> 
#> Columns 19 to 20 -0.4196 -0.1576
#>  -0.6248 -0.3506
#>  -0.3395 -0.4763
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>