RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1544 -0.4323 -0.8139 -0.8963 -0.0333 -0.0908 -0.4849 -0.8842  0.2397
#>  -0.5664  0.6444  0.8996  0.0231  0.5493  0.8260  0.7241 -0.7981 -0.3634
#>  -0.1534  0.6625  0.1637 -0.0614  0.8171 -0.5824  0.0546  0.2685  0.1007
#> 
#> Columns 10 to 18  0.5504 -0.5404  0.3691  0.7661 -0.7663 -0.7555  0.7045 -0.2706  0.4143
#>  -0.8322  0.5681 -0.6749  0.0668  0.0277 -0.2176  0.5605 -0.6823 -0.8102
#>  -0.3551  0.3585 -0.0930 -0.4243  0.5231 -0.2784  0.8605 -0.2557 -0.3123
#> 
#> Columns 19 to 20  0.4962 -0.9533
#>   0.3988 -0.3305
#>  -0.3604  0.5456
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3805  0.0292  0.0566  0.0545  0.3890  0.5417  0.4529  0.0378  0.6457
#>  -0.1732 -0.6654  0.0941  0.6789 -0.1918 -0.3876 -0.4217 -0.0589 -0.2290
#>  -0.0293  0.1491 -0.4336  0.3855 -0.4945 -0.2230 -0.0908  0.1834 -0.2521
#> 
#> Columns 10 to 18 -0.5678  0.1863  0.2168  0.1701 -0.7720  0.0656  0.5477 -0.7245 -0.0402
#>   0.1342 -0.3390  0.1102  0.0622 -0.0858  0.1517  0.4684 -0.6290 -0.6079
#>  -0.0410 -0.2921 -0.3446  0.7696 -0.3417 -0.0701 -0.2319 -0.0001 -0.2869
#> 
#> Columns 19 to 20  0.5284 -0.5095
#>  -0.3498 -0.5461
#>  -0.0070 -0.3394
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5948 -0.3493  0.6344  0.2469 -0.5895  0.5648  0.1985 -0.1646  0.0511
#>   0.1306  0.5226  0.4865  0.1797 -0.4758  0.6780 -0.2871 -0.3352  0.0935
#>  -0.0080  0.3247 -0.3474  0.3497  0.1574  0.2345 -0.4788  0.1681  0.5437
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5968 -0.0415 -0.4262  0.0994 -0.7720  0.0771 -0.3767  0.3485  0.4391
#>  -0.5734 -0.4302  0.2932  0.3798  0.0346 -0.2898  0.1909  0.3105  0.4686
#>  -0.2907 -0.3815 -0.5777  0.3286 -0.4480 -0.1246 -0.1167  0.5335  0.1448
#> 
#> Columns 10 to 18 -0.5365 -0.2453 -0.4105 -0.6122  0.0958  0.7401 -0.0786  0.2163  0.1455
#>  -0.3523  0.2440  0.2656 -0.2946  0.3467  0.6347  0.1026 -0.1143  0.4717
#>   0.1130  0.0775 -0.0321 -0.5900  0.4009  0.0825  0.5355 -0.4131  0.2879
#> 
#> Columns 19 to 20 -0.5692  0.2150
#>   0.0226 -0.8957
#>   0.1192 -0.0806
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4178 -0.2816  0.0554  0.2697 -0.0529  0.2984 -0.2717 -0.2273  0.4431
#>   0.3885  0.3536 -0.1978 -0.0935 -0.4248  0.4529  0.1973 -0.2352  0.0807
#>   0.1778  0.4247  0.0283  0.1132 -0.4146  0.2332 -0.1626 -0.0912  0.3593
#> 
#> Columns 10 to 18  0.0548 -0.5395  0.0058  0.2507 -0.3484  0.2052  0.3649 -0.6164 -0.3351
#>  -0.2486 -0.3509 -0.0928  0.2221 -0.4813 -0.0122  0.1836 -0.6294 -0.1819
#>  -0.0044  0.0120  0.2455  0.2449 -0.4178 -0.1500  0.1259 -0.2162 -0.6006
#> 
#> Columns 19 to 20  0.1509 -0.0547
#>   0.3572 -0.5741
#>  -0.1591 -0.3355
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>