RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6833 -0.6398 -0.2434  0.6959  0.8908 -0.0643  0.6343 -0.8592  0.9663
#>   0.4515  0.3969 -0.0551  0.1912  0.2983 -0.6014 -0.3551 -0.1669 -0.9421
#>   0.0356  0.0645 -0.0547  0.3506 -0.1066  0.0223  0.2396  0.2406 -0.5739
#> 
#> Columns 10 to 18 -0.0508 -0.7991  0.9252  0.8847 -0.1242 -0.4106  0.5299 -0.4277  0.1454
#>   0.2292  0.5888  0.4409  0.1281 -0.2594  0.7661 -0.7861  0.8945 -0.6249
#>  -0.6846  0.4149 -0.0700  0.0364 -0.9384  0.0080 -0.8274  0.1797 -0.5997
#> 
#> Columns 19 to 20 -0.5105 -0.2900
#>  -0.6877 -0.3684
#>   0.2060  0.3018
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2546 -0.1985  0.2609 -0.4709  0.7566  0.0394 -0.1818 -0.4177 -0.2629
#>   0.7106  0.0954 -0.0715 -0.1985 -0.2041 -0.7616 -0.1023 -0.0597 -0.7498
#>   0.1286  0.5261 -0.4592 -0.4941 -0.1530 -0.4147  0.4091 -0.5281  0.2892
#> 
#> Columns 10 to 18 -0.3651  0.0772  0.0520  0.7239 -0.4650 -0.3078 -0.2490  0.6756  0.2274
#>   0.3558 -0.3807 -0.1990 -0.1703  0.0505  0.3360 -0.5546 -0.2459 -0.5480
#>   0.5299 -0.3114  0.2011  0.6471 -0.3215  0.0873 -0.3136  0.1366 -0.3571
#> 
#> Columns 19 to 20 -0.1951 -0.1424
#>   0.2683 -0.3532
#>  -0.4112  0.0540
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0299 -0.1336  0.4898 -0.3154  0.3949 -0.2735  0.6237 -0.3979 -0.0452
#>   0.6078 -0.0397  0.0032  0.2357 -0.1793 -0.3689 -0.4330  0.4874 -0.5762
#>   0.5023  0.2810  0.3521  0.0043  0.4564 -0.3056 -0.0976 -0.0785 -0.4293
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5085  0.2603  0.5820  0.0891 -0.7350 -0.4109 -0.0190  0.8888 -0.6725
#>  -0.1052  0.2640 -0.0217  0.1868 -0.1267  0.2378 -0.4364  0.0221 -0.5389
#>   0.6301 -0.7941  0.2142 -0.3948  0.9097 -0.1122  0.6176 -0.7466  0.5259
#> 
#> Columns 10 to 18 -0.2776  0.2571 -0.3399  0.6069 -0.6402 -0.6061  0.5812  0.8922 -0.1485
#>  -0.2340 -0.2766 -0.2500 -0.3892 -0.1564  0.0678 -0.5318 -0.3865  0.3445
#>   0.6081 -0.4503  0.7213  0.8400  0.0449  0.8756  0.2586 -0.8797  0.1008
#> 
#> Columns 19 to 20 -0.1728 -0.8637
#>  -0.0036 -0.4488
#>   0.2150 -0.0632
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2258  0.3387  0.2050 -0.5618  0.3488 -0.3609  0.6235 -0.6447  0.3475
#>  -0.3217  0.4624  0.0219 -0.1006 -0.1907 -0.4323 -0.2627  0.3440 -0.1183
#>   0.7105 -0.5887  0.5153 -0.3719 -0.2777 -0.5634 -0.2602  0.5287 -0.6496
#> 
#> Columns 10 to 18  0.0175  0.2227  0.3804  0.3871  0.1802  0.2552 -0.3987  0.2119 -0.4703
#>   0.1300 -0.1652  0.0478  0.5173 -0.6184 -0.5515 -0.4064  0.2585  0.0416
#>  -0.3089 -0.1979 -0.2038  0.3394 -0.5500  0.0410 -0.3167  0.4691 -0.0760
#> 
#> Columns 19 to 20 -0.2011  0.4112
#>   0.2142  0.2291
#>   0.2797 -0.1061
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>