RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0976  0.3290  0.1733  0.5340 -0.6455 -0.2532 -0.4060 -0.6656 -0.5434
#>  -0.4027 -0.5532 -0.5351 -0.5749 -0.1992 -0.1909  0.7491  0.0902 -0.3615
#>   0.5456  0.1635  0.4260 -0.1481  0.2184  0.5299  0.6673 -0.5646  0.5478
#> 
#> Columns 10 to 18  0.1253 -0.1625 -0.1738 -0.3715 -0.1055  0.6028 -0.2714  0.7056  0.1107
#>   0.7717  0.8124 -0.4423  0.6619  0.2449 -0.4178  0.4823  0.2609 -0.6075
#>   0.0124  0.2123  0.4166 -0.1788 -0.1413  0.1685  0.1997 -0.1577  0.1999
#> 
#> Columns 19 to 20 -0.1785 -0.0351
#>   0.0431  0.5355
#>   0.3725  0.2768
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4389 -0.4362 -0.0021 -0.5649 -0.1707  0.3586  0.5657  0.3415 -0.5106
#>   0.1878  0.0598  0.1875  0.3024 -0.0598 -0.3662  0.2029 -0.3679 -0.4624
#>   0.1270 -0.2263  0.5392  0.3202 -0.1516  0.3501  0.3151 -0.0133 -0.1870
#> 
#> Columns 10 to 18 -0.0514 -0.0643  0.1976  0.5870  0.1243  0.2253  0.4441  0.0049 -0.4595
#>  -0.1809 -0.6389 -0.5393  0.1251  0.5058 -0.1038  0.4775  0.5849 -0.0588
#>   0.1402  0.3425 -0.0905 -0.0830  0.3359  0.1733  0.3153 -0.3692  0.0543
#> 
#> Columns 19 to 20 -0.5080  0.3833
#>  -0.1659 -0.3918
#>  -0.0020  0.7568
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2114  0.2076  0.3625 -0.1094 -0.4044 -0.2682  0.5343 -0.0204 -0.2736
#>  -0.1828 -0.1189  0.4290 -0.1392  0.3913  0.1166  0.0826  0.3500 -0.2626
#>  -0.1382 -0.4054  0.1745 -0.1285 -0.1663  0.0689  0.1877  0.1318 -0.4884
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5810  0.1313  0.4483  0.1219 -0.7379  0.2389 -0.3978 -0.5547 -0.4155
#>   0.5556 -0.6203  0.2265 -0.5307 -0.4980  0.1342 -0.6755 -0.0359  0.1885
#>   0.4194  0.0821  0.0411 -0.1116 -0.1121  0.6198 -0.0359 -0.4939  0.4038
#> 
#> Columns 10 to 18 -0.9493  0.2548  0.4890  0.7636  0.1075 -0.3926  0.4928  0.5452  0.5593
#>  -0.4279  0.7532  0.3755 -0.4721 -0.0146 -0.8644 -0.4759 -0.5193 -0.0769
#>  -0.2969  0.1715  0.7556  0.8497  0.4189 -0.4702  0.5902  0.1385  0.4856
#> 
#> Columns 19 to 20 -0.4141 -0.7182
#>  -0.8011  0.3725
#>  -0.1068 -0.5102
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0062 -0.3526  0.3526 -0.2691  0.0535  0.0206  0.6538  0.5829 -0.2448
#>   0.3171 -0.1151  0.0759  0.1078  0.2994 -0.1894 -0.4802 -0.2076 -0.2297
#>  -0.0692 -0.4010  0.3638  0.0391  0.0415  0.0978  0.3095  0.1339 -0.5062
#> 
#> Columns 10 to 18 -0.0810 -0.1737  0.0407  0.1651  0.1246  0.4262  0.5768 -0.4008  0.1313
#>  -0.3000 -0.6142 -0.2869  0.1374  0.3556 -0.0323  0.5601  0.1853  0.4205
#>  -0.2064 -0.5161  0.3215  0.4138  0.0557 -0.0810  0.7895 -0.3432  0.0024
#> 
#> Columns 19 to 20  0.4570  0.6081
#>  -0.1471 -0.0730
#>   0.0785  0.6414
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>