RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0551  0.3169  0.6168 -0.2798  0.0032 -0.1811 -0.5634 -0.5439 -0.4893
#>  -0.1375  0.7960 -0.7006  0.5039  0.2082 -0.6557 -0.4833  0.5534  0.5847
#>   0.4911  0.6321  0.1413 -0.6104  0.5523 -0.1202  0.6101 -0.2821 -0.0564
#> 
#> Columns 10 to 18  0.5938  0.6437 -0.5447  0.1381  0.4654  0.3722 -0.2364 -0.2132 -0.0198
#>  -0.6090 -0.5746  0.2173  0.1759  0.1164 -0.8104 -0.7530  0.2029  0.5288
#>   0.3528 -0.1652  0.6445  0.8803 -0.2059 -0.7228  0.3633  0.5379  0.1575
#> 
#> Columns 19 to 20 -0.3019  0.7275
#>   0.7901  0.7843
#>   0.7342 -0.0390
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2612  0.2080  0.5448  0.1706  0.0517 -0.0275  0.2472  0.1270  0.2704
#>  -0.0383 -0.6329  0.1286  0.3594 -0.2045  0.6738  0.1201  0.6666  0.1648
#>  -0.2459 -0.3945  0.1672  0.0352 -0.1218 -0.1997 -0.0213  0.1045 -0.4683
#> 
#> Columns 10 to 18 -0.1937 -0.3011  0.4843  0.2394  0.2571  0.2532 -0.2947  0.0630  0.3608
#>   0.1398 -0.4441  0.5924  0.1160  0.1944  0.0604  0.7111  0.4710 -0.1815
#>  -0.1983  0.5141  0.2143 -0.0665 -0.2974 -0.3820  0.0962 -0.0029 -0.0994
#> 
#> Columns 19 to 20  0.3334  0.7130
#>   0.4631  0.4782
#>  -0.6769  0.5158
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0337 -0.6472  0.1312  0.4178  0.2260  0.4368 -0.0897  0.1978  0.0083
#>   0.1160 -0.1386 -0.5960  0.6137 -0.1114  0.2804  0.0403  0.0553  0.1350
#>  -0.1320 -0.0602 -0.4810  0.2372  0.4392  0.5052  0.1296  0.0168  0.1953
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1636 -0.1159  0.2744  0.2268 -0.3234  0.4918  0.5067  0.6081 -0.0509
#>  -0.1791  0.4516  0.8371  0.3287  0.5180 -0.1131 -0.1974  0.2262  0.3442
#>  -0.4107  0.9265  0.7500 -0.1635 -0.1411 -0.4449 -0.7547 -0.5345 -0.9227
#> 
#> Columns 10 to 18 -0.0074 -0.7405 -0.1427 -0.8217 -0.2005 -0.0188 -0.5048 -0.3896 -0.1856
#>   0.7537  0.4740 -0.4158 -0.7146  0.4077  0.1424  0.6729 -0.0219 -0.1186
#>   0.4016 -0.6992  0.7116  0.6531  0.0373  0.5880  0.1130  0.5601 -0.6382
#> 
#> Columns 19 to 20  0.6090  0.0747
#>  -0.0309  0.1484
#>  -0.6340  0.0050
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1439 -0.3013 -0.6323  0.3304 -0.0820 -0.2059  0.2487  0.2176  0.1887
#>  -0.0412 -0.5949  0.0455  0.1619 -0.1801 -0.2250  0.0982 -0.2675  0.0639
#>  -0.0221 -0.1634 -0.2781  0.5640 -0.4229 -0.0269 -0.0809  0.1661  0.2005
#> 
#> Columns 10 to 18  0.5643  0.2435  0.0883 -0.3238 -0.3508 -0.4885  0.5342 -0.1143 -0.1421
#>   0.1307  0.2672 -0.3407 -0.3869 -0.0791 -0.2261  0.6281  0.7675 -0.5874
#>  -0.5344 -0.0297  0.1830 -0.0931  0.5950  0.1214  0.0322  0.4877 -0.3529
#> 
#> Columns 19 to 20  0.0560  0.3138
#>   0.1550  0.1858
#>   0.0806  0.2358
#> [ CPUFloatType{2,3,20} ]
#>