RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9136  0.7881  0.4157 -0.0402  0.2419 -0.6484  0.1598  0.2866  0.5886
#>  -0.5752  0.8396 -0.1122  0.0963  0.8455  0.7182  0.6703  0.3840 -0.2685
#>  -0.4549  0.8393  0.5608  0.0449  0.2857 -0.3044 -0.7218  0.3904 -0.1389
#> 
#> Columns 10 to 18  0.7406  0.1316  0.0984  0.3027 -0.0399  0.9861 -0.4740 -0.2257 -0.3271
#>  -0.3913 -0.4410 -0.0932  0.1491  0.0751  0.6362  0.0890 -0.1094 -0.1188
#>   0.3502  0.2760  0.7870 -0.4984  0.1194 -0.0040 -0.2810  0.2038  0.5063
#> 
#> Columns 19 to 20  0.0162  0.9210
#>   0.0543 -0.1324
#>  -0.8710  0.4369
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6113  0.7470  0.2708 -0.0136  0.8150  0.2966  0.4634 -0.5799  0.5063
#>  -0.2852  0.7289  0.0985 -0.0306  0.5849  0.1321 -0.1594  0.0298  0.5106
#>   0.3021  0.7256  0.0118 -0.1217 -0.0146  0.1788 -0.0072 -0.1123  0.1986
#> 
#> Columns 10 to 18 -0.4104  0.2534 -0.0636  0.1670 -0.4698  0.2435 -0.2518  0.3925 -0.1423
#>  -0.1488 -0.5512 -0.0785 -0.0352 -0.3993  0.4901 -0.1482 -0.4157  0.1394
#>   0.3776 -0.0768 -0.5435  0.3707  0.0966 -0.3562  0.0845  0.0966  0.8245
#> 
#> Columns 19 to 20 -0.5522  0.6510
#>  -0.5117  0.7615
#>  -0.6460  0.1253
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5392 -0.1125 -0.0706 -0.2024 -0.2303 -0.4208  0.2712 -0.2031  0.5600
#>   0.2559  0.7971  0.1104  0.2362  0.7298 -0.5443  0.0603 -0.2605  0.7140
#>   0.1705  0.4256 -0.3058  0.3257  0.3278  0.3734 -0.3115 -0.3219  0.3132
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2448 -0.2273 -0.4834  0.0999 -0.5040 -0.4635 -0.3383  0.3183 -0.5309
#>   0.2863 -0.0561  0.4397  0.3618 -0.2993 -0.3725 -0.5086  0.3347 -0.3416
#>   0.0597 -0.4149  0.4606  0.5878  0.2573  0.1476 -0.4381 -0.1319  0.1765
#> 
#> Columns 10 to 18 -0.8579 -0.5591  0.1654  0.0498  0.3380  0.0225 -0.5729 -0.5236 -0.4053
#>  -0.5446 -0.2371  0.1486  0.2856  0.0709  0.1302 -0.1320 -0.4855  0.0341
#>  -0.7634  0.1287  0.3442  0.4152  0.5056 -0.1296  0.2596 -0.1660  0.0103
#> 
#> Columns 19 to 20  0.5226 -0.0416
#>   0.5229 -0.1488
#>  -0.4279 -0.5425
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0789  0.5765 -0.3178  0.4997  0.6624 -0.3477 -0.0867 -0.5412  0.2791
#>  -0.0373  0.5218 -0.0025  0.0981  0.6139 -0.0008  0.3319 -0.2867  0.2618
#>  -0.0506  0.7441 -0.0478  0.2578  0.5609  0.0879  0.1457  0.0757  0.5508
#> 
#> Columns 10 to 18 -0.0054  0.2708  0.4545  0.1238 -0.2822  0.6852 -0.3645  0.2225  0.1506
#>   0.2673 -0.0130 -0.2384  0.1881 -0.1243  0.5850 -0.1907  0.5253  0.4479
#>   0.1949 -0.2276 -0.6213  0.3405 -0.0617  0.0695 -0.0193  0.1358  0.3995
#> 
#> Columns 19 to 20 -0.2679  0.6506
#>  -0.3748  0.6324
#>  -0.6352  0.0833
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>