RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3976 -0.7747 -0.0885  0.2227  0.7754 -0.6161 -0.6094 -0.3512 -0.0944
#>   0.6706 -0.6940 -0.8323 -0.5208 -0.5927 -0.2102  0.7402  0.7782  0.5250
#>  -0.5168  0.2067  0.4743  0.2842 -0.0783  0.5238 -0.6283 -0.3984 -0.3125
#> 
#> Columns 10 to 18  0.6064 -0.8313 -0.4746  0.0816  0.2754 -0.6105 -0.7443 -0.5519 -0.7033
#>  -0.8651  0.1707  0.7286 -0.7927 -0.2487  0.4174  0.0821 -0.7591  0.9720
#>   0.7218  0.3842 -0.2471  0.7004  0.1137  0.2265  0.1308  0.4895 -0.8460
#> 
#> Columns 19 to 20 -0.8546  0.2409
#>  -0.6613  0.2485
#>   0.6874 -0.5938
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5727 -0.0820  0.1254 -0.1479 -0.4349  0.5431  0.2015  0.1302  0.1453
#>   0.1154  0.2840 -0.0581  0.1591 -0.0129  0.1404  0.3446  0.2182  0.1595
#>  -0.8319  0.0641 -0.4004 -0.2535  0.4423  0.0648 -0.0442 -0.3514 -0.0700
#> 
#> Columns 10 to 18  0.0157  0.3135  0.5705  0.6597 -0.3571  0.4687 -0.1320  0.5755  0.2353
#>   0.5900 -0.0173 -0.1113  0.1712 -0.2374  0.3062 -0.5328  0.1203 -0.1947
#>   0.2862  0.2283 -0.1246 -0.1401  0.2179 -0.1631  0.3703 -0.2313  0.1582
#> 
#> Columns 19 to 20  0.3740 -0.3013
#>   0.2823 -0.5305
#>   0.4310 -0.1682
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3134 -0.2964 -0.5703  0.1086  0.6412  0.1712 -0.0006  0.1654 -0.2799
#>  -0.4651 -0.2990 -0.0591 -0.1978  0.3982  0.1320 -0.2018 -0.1117 -0.1458
#>   0.2341 -0.0372 -0.3297 -0.0629  0.1803  0.4355 -0.2913 -0.3530 -0.0159
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2549 -0.3119 -0.5485 -0.3653  0.2754 -0.1116 -0.3231  0.4155  0.5272
#>   0.3854 -0.6407 -0.1874 -0.5865 -0.2252 -0.6176 -0.3336 -0.1175 -0.0838
#>   0.1270 -0.4963  0.2896 -0.0832  0.3138 -0.7109 -0.0873 -0.2985 -0.3079
#> 
#> Columns 10 to 18 -0.4825  0.0531 -0.6713 -0.3214  0.0389  0.6367  0.9063  0.5266 -0.1169
#>   0.3562  0.5150 -0.5273 -0.2879 -0.0120 -0.1637  0.4656 -0.4219 -0.4222
#>  -0.1641  0.4985 -0.1172  0.3930  0.2040  0.3834 -0.1125  0.0962  0.0816
#> 
#> Columns 19 to 20  0.5498  0.0989
#>   0.2810  0.3061
#>  -0.0516  0.5760
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5054 -0.2461 -0.6317 -0.3630  0.1661  0.6377  0.1225 -0.0591 -0.2667
#>  -0.6895 -0.4029 -0.5821 -0.6715  0.7279  0.0310 -0.0052  0.2637 -0.1043
#>   0.1746  0.0104 -0.2663 -0.3700  0.0666  0.0586 -0.1949 -0.0951  0.1107
#> 
#> Columns 10 to 18  0.1381  0.1030 -0.2059  0.1509 -0.2973 -0.0773  0.3890 -0.0524 -0.1669
#>   0.4948  0.1343 -0.3998  0.2050 -0.2784 -0.2953  0.0359 -0.2917 -0.0224
#>   0.2591 -0.3627 -0.2806  0.1993 -0.2653  0.1494 -0.3523 -0.4225 -0.1756
#> 
#> Columns 19 to 20  0.2977 -0.3253
#>  -0.0239 -0.1807
#>   0.1295 -0.4280
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>