RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0247 -0.2905  0.7413 -0.6997 -0.6083 -0.0206 -0.4095  0.2797  0.4865
#>   0.4756 -0.0725  0.4786  0.1151  0.3750 -0.9025 -0.0609 -0.4356  0.9238
#>   0.6255 -0.6676 -0.4891 -0.3211 -0.1540  0.3315 -0.2667  0.5151  0.7736
#> 
#> Columns 10 to 18  0.3009 -0.0764  0.2742  0.6873 -0.6382  0.2962 -0.0928 -0.6796  0.7600
#>  -0.6080  0.1822  0.6365  0.2459  0.1081 -0.5628  0.8254 -0.6451 -0.3492
#>   0.0653 -0.2965 -0.5402 -0.0622 -0.0608 -0.5160  0.2802  0.3629 -0.0907
#> 
#> Columns 19 to 20  0.7720 -0.5402
#>  -0.6547  0.5577
#>  -0.5051 -0.0347
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0677  0.0813  0.1652 -0.3677 -0.2444 -0.0716  0.2023 -0.0221  0.7016
#>  -0.5121 -0.0733  0.6387 -0.4156  0.1504 -0.3842 -0.1711  0.0277 -0.0120
#>  -0.1145 -0.6548  0.2937 -0.4691 -0.1733 -0.1783  0.0623 -0.2104  0.1041
#> 
#> Columns 10 to 18 -0.2963 -0.2176  0.2111  0.0378  0.1312 -0.1397  0.1895  0.0800  0.3799
#>   0.0733  0.4711 -0.3003  0.3228  0.1699 -0.1433  0.1170 -0.4037 -0.5892
#>   0.0707 -0.2452 -0.3173  0.1319  0.0976 -0.0662 -0.5592  0.0376  0.0292
#> 
#> Columns 19 to 20  0.0395 -0.1336
#>  -0.7633 -0.2204
#>  -0.2966 -0.1657
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3644 -0.1328 -0.5539 -0.5656 -0.2496  0.2788  0.1733 -0.2165 -0.1695
#>   0.0735 -0.5721  0.7772 -0.3123  0.4834 -0.6462  0.2046 -0.3766  0.1939
#>  -0.3582 -0.4914 -0.2370 -0.7496 -0.2658 -0.2747  0.3014  0.0161 -0.5304
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3711  0.6749  0.1739  0.0855 -0.7848  0.7535  0.1265  0.0702 -0.5317
#>  -0.3721  0.3986  0.4540 -0.0019 -0.4742 -0.1696  0.0722  0.7159  0.1406
#>  -0.5552  0.7373 -0.3194 -0.0002 -0.6310 -0.1662  0.0262 -0.0884 -0.1339
#> 
#> Columns 10 to 18 -0.5976  0.0053  0.2902  0.1259  0.2351  0.2872  0.3112  0.7343  0.5989
#>  -0.5662 -0.2508  0.5327 -0.4958 -0.8505  0.2495  0.1073 -0.0617 -0.1530
#>  -0.6612  0.3248  0.6170 -0.1784 -0.2295 -0.6741  0.3271  0.3456  0.1495
#> 
#> Columns 19 to 20 -0.2697 -0.2382
#>   0.5316 -0.4045
#>   0.4737 -0.1492
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3166  0.0656  0.4839 -0.2461  0.1504 -0.5444  0.4079 -0.2206  0.1870
#>   0.3345 -0.4180  0.5148 -0.3562  0.2069 -0.5013  0.5276  0.0872  0.0788
#>   0.2726 -0.3454  0.7749 -0.5947  0.1968 -0.3812  0.2364 -0.2165  0.2748
#> 
#> Columns 10 to 18  0.1506  0.1738  0.0205  0.5862  0.0223  0.2760  0.3648 -0.4001 -0.1502
#>  -0.2770 -0.4189  0.0843 -0.4857 -0.2141  0.2487 -0.0511  0.0814  0.3321
#>  -0.0769 -0.0185 -0.0837 -0.0845  0.1232  0.4694 -0.0805 -0.1628 -0.2685
#> 
#> Columns 19 to 20 -0.5298 -0.1516
#>  -0.5281 -0.3437
#>  -0.6947 -0.1017
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>