RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3545  0.2700 -0.3759 -0.2515  0.9001  0.4653 -0.1363  0.3233 -0.6411
#>  -0.8833  0.0059 -0.1707  0.1957  0.2967  0.2823  0.4128  0.4030 -0.5726
#>  -0.0273  0.5970 -0.5207  0.4115 -0.8370  0.6834 -0.4582 -0.2367 -0.8524
#> 
#> Columns 10 to 18 -0.2223  0.7122 -0.0642 -0.6847  0.6478 -0.3739  0.3450 -0.6391  0.7367
#>  -0.4654 -0.1649  0.3350 -0.6819  0.2598 -0.7748 -0.1999  0.2533  0.4468
#>  -0.7131  0.1240  0.3152 -0.4269  0.0889  0.0927  0.0955  0.4835 -0.6688
#> 
#> Columns 19 to 20  0.0775 -0.7351
#>  -0.1198 -0.2613
#>   0.0963 -0.4406
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2358 -0.2489 -0.0151  0.2039  0.3933 -0.4085  0.3037 -0.1244  0.2674
#>  -0.2193  0.3496  0.1786  0.2921  0.2650  0.0956 -0.0452 -0.2635 -0.0517
#>   0.2954  0.0693 -0.1965 -0.5758  0.3100 -0.0246  0.1357  0.5075 -0.4821
#> 
#> Columns 10 to 18 -0.5474  0.4362 -0.1647 -0.3975  0.3976 -0.2263 -0.1132  0.0846 -0.4014
#>   0.0585 -0.0991  0.0988 -0.1919  0.2260 -0.1813 -0.4000  0.2900 -0.1343
#>   0.5246  0.0337 -0.1376 -0.3898  0.4328 -0.2980 -0.0123  0.3872 -0.2202
#> 
#> Columns 19 to 20  0.0460 -0.0991
#>   0.1771 -0.0014
#>  -0.5065  0.0608
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0726  0.3047 -0.4269  0.1375  0.1393 -0.0871  0.0440 -0.2131  0.0512
#>   0.0837  0.1278 -0.3446 -0.1568  0.5297 -0.2226 -0.0360 -0.1049 -0.2847
#>   0.1432  0.3065 -0.5535  0.1047  0.1339  0.4455 -0.4683 -0.0578 -0.4441
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5693 -0.2271  0.0957  0.1568  0.2727 -0.3613 -0.4966 -0.0854 -0.4134
#>   0.6605  0.4433 -0.3228  0.5574 -0.2889  0.1573 -0.0605 -0.4846 -0.3654
#>  -0.2571  0.6066  0.5152 -0.0095 -0.4676  0.9089  0.2478 -0.8033  0.6346
#> 
#> Columns 10 to 18  0.4835 -0.1941 -0.0261  0.1550 -0.5897 -0.1157 -0.5631  0.6250  0.0118
#>   0.4459 -0.2687  0.5865  0.7345  0.4157 -0.5356  0.1885 -0.1014 -0.0826
#>  -0.5771  0.5255  0.0669  0.0506  0.8380  0.4599 -0.2760 -0.9636 -0.7627
#> 
#> Columns 19 to 20 -0.0728 -0.4217
#>   0.0068 -0.3190
#>  -0.5084 -0.0815
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0837  0.1359 -0.1945  0.0961  0.2789  0.2190  0.0676 -0.5813  0.0450
#>  -0.1900 -0.1100 -0.3118 -0.3845  0.3857 -0.2892  0.0382  0.5171 -0.0510
#>  -0.3984 -0.0807 -0.5218  0.2777 -0.0847 -0.2470 -0.0903  0.1511 -0.6470
#> 
#> Columns 10 to 18 -0.1406  0.1921 -0.3322 -0.3504 -0.2093 -0.0256  0.0379 -0.0795 -0.1696
#>  -0.0596  0.2919 -0.3263 -0.4369  0.3276 -0.1179  0.0860  0.1048 -0.1240
#>  -0.3586  0.0719  0.1553 -0.3634  0.6860 -0.5784 -0.1437  0.5385 -0.5164
#> 
#> Columns 19 to 20 -0.1797 -0.1326
#>  -0.2069 -0.0843
#>  -0.0474 -0.0530
#> [ CPUFloatType{2,3,20} ]
#>