RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8636  0.2800  0.5433  0.4083 -0.0799  0.4973  0.6994 -0.4978  0.7761
#>   0.8887 -0.4025 -0.8260  0.5557  0.0830 -0.1857  0.8502 -0.3576 -0.4518
#>   0.5823 -0.8910  0.0171  0.2487  0.7567 -0.4114 -0.6534 -0.1593 -0.1111
#> 
#> Columns 10 to 18 -0.5432 -0.4038  0.3131  0.0398 -0.4471 -0.6978 -0.8431 -0.1020 -0.4417
#>  -0.3575  0.8770  0.3798 -0.3410 -0.6410 -0.8563 -0.3552 -0.4815  0.0246
#>   0.0613  0.6858  0.7407 -0.5552  0.3414  0.2804 -0.2487 -0.0551 -0.1510
#> 
#> Columns 19 to 20 -0.1488  0.4855
#>  -0.5615 -0.6592
#>  -0.6895 -0.1941
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6652 -0.1146  0.1987 -0.5695  0.1465  0.5147  0.5602 -0.1939 -0.4378
#>   0.2974 -0.6365 -0.2128 -0.1070 -0.6909  0.2788  0.3582 -0.2866  0.6300
#>   0.3902 -0.3093 -0.3148 -0.4015 -0.1610 -0.2910  0.2193 -0.6140  0.1678
#> 
#> Columns 10 to 18 -0.4890  0.0152 -0.0585  0.2685  0.0790  0.3424 -0.6855  0.3096 -0.5601
#>   0.2190  0.7256  0.2270  0.0145 -0.7497 -0.0515  0.0368 -0.3543 -0.0581
#>   0.3719  0.1355 -0.1306 -0.2249 -0.0764  0.0477 -0.5217  0.3342  0.3729
#> 
#> Columns 19 to 20 -0.2285 -0.2940
#>  -0.1429 -0.4888
#>  -0.0548 -0.5298
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1575 -0.4374  0.5142  0.2084  0.0149  0.1314 -0.2269 -0.2842 -0.2481
#>  -0.3769 -0.6940  0.5797 -0.0671 -0.3038  0.4717  0.5939 -0.2816  0.1279
#>  -0.0005 -0.2913  0.3483 -0.3586 -0.2952  0.1752 -0.0474  0.0029  0.1405
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0619  0.6399 -0.0376 -0.1692  0.0618 -0.0373 -0.3515 -0.1398  0.2217
#>   0.4130 -0.4762 -0.1642  0.0518 -0.2318 -0.0481 -0.2158 -0.7957  0.2544
#>   0.8163 -0.3120 -0.0946 -0.0287 -0.1876 -0.4937  0.0818 -0.2665 -0.3989
#> 
#> Columns 10 to 18 -0.2464  0.4611 -0.2643 -0.4178  0.4955 -0.0692  0.4459  0.4298 -0.5911
#>  -0.3755  0.5374  0.1842 -0.5874  0.3083 -0.0211  0.5860  0.0472  0.3822
#>  -0.2500  0.4486  0.1213 -0.5315  0.2019 -0.2206  0.5995 -0.2706  0.7321
#> 
#> Columns 19 to 20 -0.3991  0.8035
#>  -0.1797 -0.0657
#>   0.7502 -0.3935
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1464 -0.4650  0.1811 -0.1448 -0.3338  0.2647  0.0485 -0.2709  0.0988
#>  -0.2531 -0.2659  0.2144 -0.3180 -0.1530  0.1333  0.0031 -0.4696 -0.0622
#>   0.1882 -0.0692  0.1076 -0.3729 -0.0591  0.3016 -0.0287 -0.4244 -0.3649
#> 
#> Columns 10 to 18 -0.2663  0.3311  0.1805 -0.0148 -0.3373  0.1045 -0.0309  0.3343  0.0133
#>   0.0169  0.2340 -0.3442 -0.0437 -0.3536 -0.0899 -0.2882  0.0458 -0.0855
#>   0.0516  0.3953 -0.7157 -0.2204 -0.0227 -0.1407 -0.5370  0.0703  0.2496
#> 
#> Columns 19 to 20 -0.2663 -0.1711
#>  -0.0019 -0.2328
#>  -0.0464 -0.5341
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>