RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8459 -0.1389 -0.9595  0.7961 -0.3184  0.2801  0.0816  0.1444 -0.8338
#>   0.6637 -0.7641 -0.7811 -0.8650  0.0590  0.3721 -0.0211  0.7337 -0.8433
#>   0.2858 -0.3835 -0.0351 -0.0914 -0.1268 -0.7094 -0.6930 -0.7270 -0.0342
#> 
#> Columns 10 to 18  0.1219 -0.5643  0.0286  0.1154  0.7399  0.9086 -0.4439 -0.5310  0.1646
#>  -0.0144 -0.6994  0.8617 -0.3029  0.3342  0.0148  0.1310  0.1322  0.4491
#>  -0.7467  0.1040 -0.4139  0.2142  0.2114  0.0117  0.4985  0.0179 -0.3544
#> 
#> Columns 19 to 20  0.5532 -0.8906
#>  -0.2158  0.3369
#>  -0.3788  0.4992
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0689  0.4039 -0.5226  0.1421  0.4101 -0.3012 -0.5542  0.2290 -0.2596
#>   0.2213 -0.6363  0.2511 -0.3000 -0.8087 -0.2197 -0.2275  0.2114 -0.3229
#>   0.3748 -0.3698 -0.2385 -0.2950 -0.0971 -0.0612  0.1286  0.5358 -0.3129
#> 
#> Columns 10 to 18 -0.6903 -0.2660  0.3852 -0.2528  0.5221 -0.4391 -0.1439 -0.0109  0.1814
#>  -0.3071 -0.3014  0.2520 -0.3285 -0.5212  0.0938 -0.1135  0.5435  0.2113
#>   0.1775  0.0411  0.1523 -0.2556  0.5159  0.7623 -0.1948 -0.5697 -0.0307
#> 
#> Columns 19 to 20 -0.0076  0.3161
#>  -0.3578  0.0017
#>  -0.2819 -0.2230
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3795 -0.2737  0.1804 -0.7113 -0.4307  0.4123 -0.5042 -0.1742  0.2789
#>   0.4746 -0.5787  0.1933 -0.3106 -0.2569 -0.0448 -0.1884  0.1507 -0.4863
#>   0.3116 -0.2925 -0.2719 -0.5932  0.2499 -0.0269 -0.5381  0.0950 -0.1382
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1697 -0.1154  0.5831 -0.5155  0.3259  0.7534 -0.0043  0.0534 -0.5911
#>  -0.1151  0.5446  0.3745 -0.1589  0.5005  0.4076  0.3569  0.1984 -0.2689
#>  -0.4602 -0.4390 -0.0064 -0.2736  0.2976  0.7573 -0.2873 -0.2545 -0.5671
#> 
#> Columns 10 to 18 -0.2836  0.8035  0.0508 -0.0613  0.0097  0.3328  0.0989 -0.1360  0.0778
#>   0.1984  0.4191  0.3299 -0.6310  0.3619  0.3716  0.1881 -0.4910 -0.5330
#>  -0.2165  0.7930  0.0514  0.2984 -0.2339  0.5304  0.5431 -0.8951 -0.0780
#> 
#> Columns 19 to 20  0.2159  0.4493
#>   0.1976 -0.1995
#>  -0.0367  0.7559
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2971 -0.4188  0.3246 -0.6847 -0.1689 -0.0249  0.0478  0.1616  0.0415
#>   0.5180 -0.5840 -0.0032 -0.4031 -0.1462  0.1366 -0.1801  0.3343 -0.1974
#>   0.5046 -0.1458 -0.0007 -0.7102 -0.1414  0.1273 -0.2034  0.1679 -0.1731
#> 
#> Columns 10 to 18 -0.0949 -0.4640  0.4394  0.1545 -0.3634  0.1533 -0.3788 -0.0379 -0.3746
#>  -0.3473 -0.5888  0.2731 -0.3551 -0.1007  0.0720  0.0170  0.4635  0.5009
#>  -0.3810 -0.5952  0.2298  0.1987 -0.3801 -0.2078 -0.6887 -0.2979 -0.4866
#> 
#> Columns 19 to 20 -0.1658 -0.3863
#>  -0.3946 -0.4039
#>   0.2531 -0.4276
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>