RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8243 -0.6081  0.8925  0.9003 -0.5751 -0.4112 -0.6588 -0.0706 -0.1475
#>  -0.0358 -0.7841  0.2579  0.6227 -0.3389 -0.0853  0.7317  0.7880  0.5789
#>  -0.1702 -0.1562 -0.7336 -0.0226 -0.1907 -0.7380 -0.3254 -0.3804  0.4895
#> 
#> Columns 10 to 18 -0.5674  0.4418  0.5470  0.6138  0.3929 -0.3843  0.5300  0.4595  0.1985
#>  -0.8431 -0.1619  0.6414  0.9744 -0.0450 -0.3579  0.0195  0.6514  0.0688
#>   0.0254 -0.9297 -0.0974 -0.2134  0.1648 -0.2993 -0.2754 -0.5325 -0.7178
#> 
#> Columns 19 to 20  0.3993  0.7268
#>   0.6910  0.5536
#>   0.6004  0.2036
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4832  0.2404 -0.6571  0.0256 -0.1812 -0.0137 -0.4617 -0.0373 -0.0039
#>  -0.2491  0.6524 -0.4952 -0.0804 -0.1762 -0.2874 -0.7279 -0.1589 -0.4686
#>  -0.1895 -0.7112 -0.0576 -0.0675 -0.3953  0.1752 -0.0750  0.4229 -0.3556
#> 
#> Columns 10 to 18  0.4383  0.2339  0.0055 -0.3303 -0.5167 -0.1658  0.3591  0.1324  0.5092
#>   0.0642  0.1687 -0.0994  0.0107 -0.4785 -0.3934 -0.1304  0.0071  0.5474
#>   0.0842 -0.1648 -0.2944  0.1312 -0.0937 -0.1007  0.1435 -0.4763 -0.4629
#> 
#> Columns 19 to 20 -0.2472 -0.8152
#>   0.3983 -0.5840
#>   0.4618 -0.3359
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2207 -0.2212 -0.3945  0.3819  0.0061 -0.2870 -0.8412  0.6998 -0.3174
#>  -0.6154  0.3413 -0.3756  0.5809 -0.0615  0.3304 -0.5566  0.7152  0.1591
#>   0.2976 -0.1128 -0.6985 -0.0456 -0.2700 -0.4032 -0.5099  0.2454 -0.0024
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8004 -0.8906 -0.2588 -0.4918  0.2912  0.2430 -0.4894  0.4186 -0.2274
#>   0.5872 -0.4599  0.3459  0.3954  0.3673  0.3516  0.2838 -0.0710 -0.8586
#>   0.5708 -0.7377 -0.0464  0.3532  0.5612 -0.4768  0.0139  0.7031 -0.4096
#> 
#> Columns 10 to 18  0.8376  0.7514 -0.7804  0.1137 -0.4371 -0.7666  0.7657 -0.1186 -0.1057
#>   0.7861 -0.5234  0.6171 -0.2056 -0.3918 -0.6862  0.4446 -0.3704  0.9010
#>   0.6358  0.4240 -0.6779  0.0669  0.4195 -0.7762  0.8878  0.3306 -0.0671
#> 
#> Columns 19 to 20  0.5187  0.1622
#>  -0.3532 -0.6622
#>   0.2364  0.1920
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7996  0.2995 -0.2149  0.6138  0.7084 -0.3878 -0.1399  0.5077 -0.0945
#>  -0.6848  0.8405 -0.4532  0.1127  0.2695 -0.5455 -0.5383  0.3453  0.0639
#>  -0.7106  0.0685 -0.5882  0.4527  0.4539 -0.3349 -0.5181  0.5640  0.0175
#> 
#> Columns 10 to 18 -0.6847 -0.0497  0.0145  0.2417 -0.5643  0.3049  0.2775 -0.3553  0.3746
#>  -0.1527  0.1285 -0.1365 -0.2593 -0.2344  0.6657 -0.2033 -0.3734  0.2641
#>  -0.6730  0.1343 -0.2987 -0.1508 -0.4172  0.4899  0.4611 -0.2009  0.2480
#> 
#> Columns 19 to 20  0.1027 -0.4309
#>  -0.2058  0.0366
#>  -0.1397 -0.3714
#> [ CPUFloatType{2,3,20} ]
#>