RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2627  0.6710 -0.7194 -0.3377  0.1943 -0.1427  0.6783 -0.1144 -0.4072
#>  -0.6314  0.5620  0.1214 -0.6008 -0.2917 -0.3625  0.3927 -0.6682  0.5183
#>  -0.7169  0.8860 -0.5710  0.0150  0.7236  0.2447  0.7313 -0.6104 -0.3224
#> 
#> Columns 10 to 18 -0.2450  0.3508 -0.1504 -0.4063 -0.5632 -0.8919 -0.1192  0.1350 -0.1206
#>   0.9144  0.5713  0.3452  0.6084 -0.0504  0.3576 -0.5876  0.3527  0.4057
#>  -0.6738  0.0671 -0.0762  0.5083 -0.6672 -0.0352 -0.0184  0.4610  0.0775
#> 
#> Columns 19 to 20  0.1253  0.0137
#>  -0.5668  0.3172
#>   0.5807  0.5279
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1240 -0.3309 -0.0784 -0.2081  0.0339 -0.3582  0.3205  0.2590 -0.7525
#>  -0.7297  0.3198 -0.1657 -0.1417  0.1323  0.1745  0.0329 -0.2176  0.3104
#>   0.0288 -0.3027 -0.0179  0.2116  0.1163  0.1373  0.1247  0.2444 -0.8058
#> 
#> Columns 10 to 18  0.6693 -0.4783 -0.0154 -0.0040 -0.1696 -0.2963 -0.2262  0.5979  0.2725
#>   0.7976  0.5715  0.2901 -0.0728 -0.2966  0.2219 -0.0212  0.0298  0.4667
#>   0.5645  0.1224  0.3655 -0.2052 -0.4443  0.0060  0.0714  0.6328 -0.0478
#> 
#> Columns 19 to 20  0.2254 -0.0569
#>  -0.1201 -0.1126
#>  -0.0275 -0.1336
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3984 -0.2633 -0.0713  0.2146  0.0210 -0.0937  0.0631  0.1553 -0.4005
#>  -0.2417 -0.2651 -0.0549 -0.7710 -0.1318  0.2913  0.1224  0.3841 -0.1973
#>  -0.4739  0.2235 -0.1611 -0.0743 -0.1846  0.0845  0.2973 -0.1738 -0.4869
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5769  0.5091 -0.3530  0.6087  0.2162 -0.0051  0.4396 -0.6112  0.0345
#>  -0.5116 -0.4191  0.5471 -0.0400 -0.5565 -0.7301 -0.2012 -0.8276  0.2090
#>   0.5974  0.1204  0.6551 -0.4382  0.6307 -0.2281 -0.7758 -0.5578  0.0491
#> 
#> Columns 10 to 18  0.8063 -0.8591  0.6177  0.2011 -0.3140 -0.2726 -0.4176 -0.5621  0.5413
#>  -0.0569  0.3693  0.0454  0.2019 -0.6583  0.6107 -0.2990  0.2006 -0.5566
#>  -0.7162  0.2678 -0.2571 -0.0280  0.0267  0.7308  0.4541 -0.2310  0.3004
#> 
#> Columns 19 to 20  0.4319 -0.3465
#>   0.5947  0.3470
#>   0.7401 -0.2344
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4886  0.2070 -0.2270  0.2757  0.2857 -0.0640  0.1385  0.0840 -0.4659
#>  -0.6337  0.3065  0.4565 -0.0062  0.0558 -0.0892  0.4331 -0.4352 -0.4037
#>  -0.4228 -0.2744  0.1126 -0.1820  0.2669  0.3639  0.0323  0.2438 -0.6657
#> 
#> Columns 10 to 18  0.5249  0.0239  0.4105  0.2150 -0.0671  0.3471  0.0395  0.5300 -0.2483
#>   0.5470  0.2313  0.0674  0.6758 -0.0329  0.3603 -0.0192  0.4370  0.4287
#>   0.0389  0.2252 -0.0941 -0.2509 -0.3493 -0.1922  0.6400  0.5522 -0.0835
#> 
#> Columns 19 to 20  0.3795 -0.1381
#>   0.4152 -0.1975
#>   0.1718  0.1677
#> [ CPUFloatType{2,3,20} ]
#>