RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7770  0.7591  0.3521 -0.6619  0.0816 -0.7881  0.0308  0.5736  0.4050
#>   0.8494 -0.2355  0.3664 -0.5483 -0.6724 -0.4130 -0.2409  0.1254  0.1077
#>   0.7107  0.2908  0.6860 -0.5563 -0.8471 -0.7486 -0.2451 -0.5087  0.0801
#> 
#> Columns 10 to 18  0.4487 -0.1913  0.3711 -0.2913  0.7883 -0.5951  0.2470  0.8314 -0.3574
#>   0.5299 -0.5110  0.2222 -0.0556  0.4360  0.7279  0.2225  0.2600  0.3597
#>   0.2301 -0.6868  0.7220  0.3744  0.9561 -0.7359  0.0016  0.5915  0.0727
#> 
#> Columns 19 to 20 -0.7051 -0.6905
#>  -0.0394 -0.5965
#>  -0.8354 -0.4778
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4077 -0.0421 -0.2492  0.4909 -0.2141  0.2432  0.3110 -0.5287  0.5524
#>   0.4615  0.3089 -0.5175 -0.1705 -0.2092 -0.5600 -0.0462 -0.0561  0.6705
#>   0.2715 -0.3657  0.1058  0.3534  0.1107  0.4549 -0.1318 -0.0676  0.0635
#> 
#> Columns 10 to 18  0.1288  0.0521 -0.4062 -0.6205 -0.5161 -0.4160  0.2277  0.6349  0.3601
#>   0.7500  0.3148 -0.3735 -0.5101 -0.7721 -0.0362 -0.3678  0.4518  0.3920
#>   0.3946 -0.3113 -0.2021 -0.2085 -0.2282 -0.2517 -0.0726  0.3247  0.0284
#> 
#> Columns 19 to 20 -0.5455  0.3160
#>  -0.0401  0.2451
#>  -0.2945  0.3450
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1133  0.6820 -0.6295 -0.2284  0.0698 -0.5464  0.0521  0.1640  0.4543
#>   0.1173  0.4159 -0.4524 -0.0744  0.6755 -0.4879  0.5164 -0.2674  0.5633
#>   0.5042  0.7256 -0.2415  0.3976  0.1923 -0.4002  0.1649 -0.2818  0.2517
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4473 -0.2074  0.3810 -0.3422 -0.1124  0.1247  0.0635  0.3467 -0.2226
#>   0.0903  0.5630 -0.2677 -0.0797  0.2798  0.3395  0.3148  0.4213  0.3171
#>  -0.2046 -0.8741  0.8176  0.3915  0.4669 -0.4244 -0.2314 -0.1642 -0.0621
#> 
#> Columns 10 to 18  0.4812 -0.0065 -0.2011 -0.3096 -0.0146 -0.2488 -0.5409  0.4044  0.5665
#>   0.8587 -0.7903  0.1400  0.4222 -0.2443  0.5128 -0.3723 -0.6705 -0.1443
#>   0.7040 -0.2096 -0.3258 -0.5471  0.1885 -0.1900  0.2048 -0.0523 -0.1994
#> 
#> Columns 19 to 20 -0.3470  0.4148
#>   0.3772 -0.4025
#>  -0.3621 -0.0093
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3675  0.0340 -0.4845 -0.0660  0.2497  0.0130  0.0365 -0.1396  0.2684
#>   0.1157  0.6609 -0.4701 -0.4428  0.1985 -0.0955  0.3900 -0.1215  0.4191
#>   0.2265  0.2519 -0.1049  0.3116  0.2778 -0.4397  0.0772  0.1856  0.5670
#> 
#> Columns 10 to 18 -0.1344 -0.3790  0.2218  0.0252 -0.0449 -0.3170 -0.1482  0.2412  0.2131
#>   0.0762  0.1584 -0.1913 -0.3378 -0.1731 -0.4078 -0.4536  0.7017  0.3368
#>   0.2874 -0.0862  0.1239 -0.2894  0.1520 -0.4319 -0.1057 -0.1126  0.2114
#> 
#> Columns 19 to 20  0.1509 -0.0952
#>  -0.1036 -0.1247
#>  -0.1012  0.1113
#> [ CPUFloatType{2,3,20} ]
#>