RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.9711  0.3191 -0.9185 -0.0679 -0.3488  0.8784  0.1025  0.4533 -0.3303
#>   0.6317  0.1871  0.6929 -0.0496 -0.2991  0.7572  0.3058 -0.2372  0.8472
#>  -0.3882 -0.8225 -0.4127  0.2559 -0.2255  0.0336 -0.7323 -0.4638 -0.2214
#> 
#> Columns 10 to 18  0.0235 -0.4625 -0.8377 -0.6857  0.7101  0.7659 -0.5984 -0.9621  0.1973
#>  -0.2483 -0.6471 -0.5039  0.3373 -0.0001  0.1503 -0.2326 -0.0816 -0.6323
#>  -0.6542 -0.6231  0.1002  0.6719 -0.0130  0.7006  0.8385 -0.5091  0.8255
#> 
#> Columns 19 to 20 -0.7438 -0.3172
#>  -0.2919 -0.1226
#>   0.4320 -0.1126
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5127 -0.5269 -0.6541  0.2824 -0.5023  0.4458 -0.4983 -0.2753  0.1018
#>   0.6951 -0.6431 -0.0969 -0.1573 -0.7467  0.1330  0.2548  0.0257  0.1381
#>   0.3609 -0.7658  0.1908 -0.3241 -0.1132 -0.6380  0.4498  0.5865  0.0502
#> 
#> Columns 10 to 18  0.1872 -0.5198  0.4259  0.5457 -0.4015  0.3115 -0.2107 -0.4457  0.5051
#>   0.3662 -0.5714  0.1690  0.0977 -0.2851 -0.2274  0.1711 -0.7928  0.1685
#>   0.2988 -0.1985  0.5220  0.5444 -0.1467 -0.2479 -0.1532 -0.6514 -0.1739
#> 
#> Columns 19 to 20  0.1338 -0.6526
#>  -0.1580 -0.5484
#>  -0.1269 -0.2567
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0248 -0.8139 -0.1210  0.0268 -0.4603 -0.1560  0.6314 -0.1397  0.1634
#>   0.2245 -0.8368 -0.0888 -0.1877 -0.5801 -0.3749  0.3414 -0.2208 -0.1525
#>  -0.0056 -0.6308  0.4135 -0.7924 -0.4163 -0.6532  0.6709  0.2346 -0.2981
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8374 -0.5150  0.2822  0.0452  0.7127  0.6025 -0.0368 -0.2096 -0.6912
#>  -0.3976 -0.6050 -0.0493 -0.3039 -0.0037 -0.0752  0.3331 -0.4273  0.3179
#>  -0.2570  0.1494  0.3584 -0.7130 -0.1444  0.3209  0.4555 -0.1258  0.2472
#> 
#> Columns 10 to 18  0.7441 -0.2391 -0.4100  0.2469  0.8589  0.8979 -0.3063 -0.7525 -0.0696
#>   0.0796 -0.2518 -0.0213 -0.4808 -0.2610  0.6989  0.2563 -0.3228  0.6261
#>  -0.6726  0.7472 -0.3006 -0.3563  0.2112  0.1719  0.5576 -0.1565  0.2125
#> 
#> Columns 19 to 20 -0.6415  0.6392
#>   0.0517 -0.4836
#>   0.3256 -0.1169
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0082 -0.6953  0.3059 -0.3504 -0.3669 -0.3935  0.1686 -0.1554 -0.2132
#>   0.4079 -0.6885  0.1655 -0.3333 -0.2247 -0.0977  0.2605  0.0185 -0.3595
#>  -0.0866 -0.5779 -0.0791 -0.4366 -0.4211  0.1216  0.3673 -0.1066 -0.4044
#> 
#> Columns 10 to 18  0.0178 -0.6337  0.3729  0.6319  0.0251  0.1136 -0.2918 -0.1597  0.1312
#>  -0.2976 -0.3852  0.3942 -0.1307 -0.0749 -0.2295 -0.5956 -0.0929 -0.0813
#>  -0.2371 -0.1778  0.4936  0.1712 -0.1835 -0.3163 -0.2340 -0.4939  0.3808
#> 
#> Columns 19 to 20 -0.0559 -0.5086
#>  -0.0348 -0.2936
#>   0.3762 -0.4708
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>