RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4534  0.5025  0.5519 -0.3158 -0.1950 -0.2848  0.7073  0.8660  0.4648
#>   0.8886  0.5873  0.5716  0.5664  0.3701  0.2818 -0.3830 -0.2285 -0.5645
#>  -0.4481  0.6362 -0.3762  0.0514  0.5472 -0.1919  0.0188  0.5465 -0.4934
#> 
#> Columns 10 to 18  0.5165  0.5112 -0.3288  0.7198  0.1830 -0.3027  0.4955  0.6093 -0.6881
#>  -0.8015  0.6600 -0.8408  0.5745  0.2842  0.3727  0.8658 -0.1290 -0.2052
#>   0.4736  0.0855 -0.6685  0.5274 -0.7211 -0.1092  0.2137 -0.5534 -0.8625
#> 
#> Columns 19 to 20 -0.2701  0.7301
#>   0.6256  0.7667
#>   0.0263 -0.3387
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1477  0.5653 -0.7075  0.5308 -0.5840 -0.3508 -0.1386  0.5968  0.3860
#>   0.0955 -0.3653  0.4683  0.6304 -0.2972 -0.4907 -0.0096  0.3805  0.0900
#>   0.1636  0.6960 -0.0242  0.2507 -0.2523 -0.6993 -0.1931  0.1681  0.0231
#> 
#> Columns 10 to 18 -0.3035 -0.2251  0.4467  0.3201 -0.3987 -0.5468 -0.4782  0.3830  0.1588
#>  -0.1570  0.2339  0.2045 -0.0327 -0.4009  0.5799 -0.3551  0.1459 -0.2095
#>   0.2237  0.2885  0.1114  0.1620 -0.4928  0.4127 -0.2933  0.3531 -0.0964
#> 
#> Columns 19 to 20 -0.0741  0.2946
#>  -0.2427  0.7815
#>   0.5271  0.1426
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3779  0.5858 -0.5616  0.3816 -0.3624 -0.0759 -0.2926  0.3568 -0.2348
#>  -0.4069  0.3072  0.0227  0.1814  0.2226 -0.1256  0.1320  0.8180 -0.1066
#>  -0.4311  0.4254 -0.7363  0.3205 -0.2934 -0.2039 -0.1009  0.4518 -0.5654
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2744 -0.8590  0.2463 -0.2609  0.2096 -0.0129  0.5145 -0.7765 -0.4854
#>   0.5267 -0.1464 -0.6525  0.1382  0.4133 -0.8809 -0.6106  0.3445 -0.5485
#>  -0.3395 -0.4962 -0.1195 -0.2417  0.2670  0.3490 -0.1990 -0.7144  0.0098
#> 
#> Columns 10 to 18  0.5180  0.2157  0.1179  0.5557  0.7730 -0.0833  0.0663 -0.3476  0.2193
#>   0.0609  0.2789  0.3408 -0.3091 -0.0129  0.7479 -0.3290  0.0019  0.3221
#>   0.0922 -0.5240  0.6439  0.3681  0.0408  0.5382  0.3508 -0.0725  0.7276
#> 
#> Columns 19 to 20 -0.4633 -0.4769
#>  -0.8432 -0.4901
#>  -0.4406 -0.6133
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1285  0.6961 -0.5337  0.4579 -0.5601 -0.1863 -0.3268  0.5466  0.1975
#>  -0.6102  0.1988 -0.6683  0.2492 -0.3146  0.2515 -0.1299  0.5018  0.1496
#>  -0.0591  0.4668 -0.4241  0.3093 -0.3646 -0.4002 -0.0014  0.3056 -0.0864
#> 
#> Columns 10 to 18 -0.1261  0.1251  0.3110  0.3087 -0.5901 -0.2455  0.1720  0.3422 -0.1139
#>  -0.3695 -0.2767  0.0971  0.3590  0.0206 -0.6905  0.0374 -0.3115 -0.3944
#>   0.0826  0.0588 -0.2986  0.1148 -0.5435  0.0699 -0.1072  0.2080 -0.3250
#> 
#> Columns 19 to 20  0.1801  0.5721
#>  -0.0756  0.3397
#>   0.0965  0.4464
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>