RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7705  0.7579  0.3903  0.0944  0.3893  0.1616 -0.2515 -0.6830  0.7899
#>  -0.0972 -0.2097 -0.3213 -0.0654  0.3812  0.8606 -0.4739 -0.2866 -0.5287
#>   0.4941  0.7268  0.2012 -0.3651  0.1555  0.0643 -0.2022 -0.2484 -0.6031
#> 
#> Columns 10 to 18 -0.7094 -0.0003 -0.1337 -0.0276  0.2352 -0.6264  0.5381  0.7484  0.0393
#>   0.4381 -0.5044 -0.3470  0.2682  0.3670 -0.1589 -0.3529 -0.1619  0.8439
#>  -0.0390 -0.0954  0.4070  0.0858  0.1639 -0.3895 -0.0940 -0.8641 -0.2856
#> 
#> Columns 19 to 20 -0.4912 -0.2171
#>   0.9086 -0.0114
#>   0.9364  0.7010
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0770  0.4215 -0.5188  0.0397 -0.2712  0.4827 -0.6574 -0.6555  0.0994
#>  -0.3451 -0.2942  0.2572 -0.0485  0.4963  0.0868 -0.2236  0.2611  0.0566
#>   0.0391 -0.2804  0.6041  0.1297  0.1922 -0.0060 -0.3557  0.1071  0.1642
#> 
#> Columns 10 to 18  0.5082 -0.2300  0.6013  0.1618 -0.1583 -0.4192  0.0531  0.1008 -0.2463
#>   0.0655 -0.0370  0.2697  0.3084  0.0855 -0.7487  0.4517 -0.2512  0.0162
#>   0.4119 -0.0384  0.4583  0.5686  0.4040 -0.6979 -0.0824 -0.3483 -0.3579
#> 
#> Columns 19 to 20 -0.3357  0.3819
#>  -0.3113  0.3953
#>   0.1055  0.6394
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1635 -0.2926  0.0019 -0.0068 -0.0809  0.0984 -0.0583  0.0002  0.3604
#>   0.3596  0.5033 -0.0669  0.2519 -0.0931  0.2265  0.1003 -0.1219  0.3697
#>   0.3234  0.0893  0.6340  0.1320 -0.0312  0.0790 -0.1817  0.0999  0.7315
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2990 -0.3221 -0.3015 -0.5448 -0.2166 -0.3848 -0.2139  0.7073 -0.3851
#>   0.2556 -0.1063 -0.2934 -0.7020  0.0489 -0.0275 -0.4167  0.9036  0.0186
#>   0.7439  0.3157  0.6944 -0.5722  0.5908  0.0037  0.5923 -0.0437 -0.1783
#> 
#> Columns 10 to 18 -0.1949  0.2053 -0.1400 -0.2849 -0.1774 -0.8305  0.2119 -0.3334 -0.2452
#>   0.5180  0.2611 -0.4835 -0.2304 -0.6757 -0.7488  0.6519 -0.4289  0.4419
#>  -0.3099  0.5341  0.0405 -0.1178  0.3560 -0.5105 -0.7067  0.2332  0.2348
#> 
#> Columns 19 to 20 -0.2412 -0.5819
#>   0.5918 -0.7086
#>   0.5984  0.5236
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1374 -0.3456  0.5534 -0.0974  0.2787  0.2166 -0.0409 -0.2946  0.6591
#>   0.6103  0.1835 -0.0427  0.0380  0.4246  0.0149  0.1204 -0.3327  0.4456
#>  -0.3172 -0.5138  0.4534 -0.0501  0.1030 -0.0640 -0.7408  0.0176  0.0616
#> 
#> Columns 10 to 18  0.4931 -0.0659  0.3029  0.5143  0.3910 -0.6426  0.6536 -0.2006  0.1540
#>   0.5654 -0.0062  0.7920  0.1261  0.5082 -0.2615  0.6119 -0.1258 -0.3240
#>   0.1561 -0.3723 -0.2096  0.1256  0.0862 -0.4631  0.3411 -0.0686 -0.0210
#> 
#> Columns 19 to 20 -0.1865 -0.0321
#>  -0.0112 -0.1897
#>  -0.0525  0.6021
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>