RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3050  0.4764  0.7907  0.1812 -0.2160  0.7275  0.8117  0.6488 -0.5652
#>  -0.9830  0.3163 -0.2659  0.7654  0.4705 -0.7398  0.4925  0.4683  0.2635
#>  -0.3097  0.8084  0.8515  0.1917 -0.4856 -0.3987  0.5427  0.5088  0.8370
#> 
#> Columns 10 to 18 -0.7389 -0.5798 -0.3169 -0.3016  0.7595 -0.4147 -0.4670 -0.5912  0.3599
#>   0.2846  0.9583  0.4425  0.2477 -0.7232  0.5165  0.2005  0.7904  0.5820
#>   0.8302  0.9259 -0.4323  0.8050  0.2866 -0.0303  0.0170  0.5104  0.2407
#> 
#> Columns 19 to 20 -0.1013  0.4964
#>   0.6380 -0.6889
#>   0.2653 -0.6958
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4691  0.0193  0.0235  0.6211  0.2425  0.5662  0.0793  0.5002 -0.5316
#>  -0.2727  0.4910  0.5983  0.0175  0.6416 -0.6306  0.6649 -0.1100 -0.1315
#>   0.0155  0.7036  0.6776  0.3384 -0.1424  0.2637  0.1780  0.8060  0.6162
#> 
#> Columns 10 to 18 -0.3366 -0.7992 -0.5002  0.2926  0.0804 -0.4325 -0.2720 -0.7797 -0.3378
#>   0.0576  0.4753 -0.0271 -0.7251 -0.2352  0.5242  0.7479 -0.2108  0.0130
#>   0.1702  0.2729 -0.1344 -0.1156  0.6500  0.0908  0.3475 -0.4973  0.1194
#> 
#> Columns 19 to 20  0.5454 -0.3610
#>   0.3557  0.1435
#>   0.3158  0.3290
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2274  0.1519  0.2662  0.0263 -0.3926  0.4630  0.4426  0.4844 -0.2426
#>  -0.0749 -0.2056  0.3154 -0.1087  0.4660  0.4134 -0.1051  0.1146 -0.0115
#>  -0.1471  0.3608  0.4134  0.5809 -0.0666  0.3372  0.0822  0.5034 -0.0566
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3041  0.2724 -0.0730  0.2184 -0.2779  0.1338 -0.0202 -0.7642  0.4044
#>  -0.6689  0.3283 -0.3511  0.1745  0.2232  0.2664 -0.6302  0.4496 -0.6113
#>  -0.1721 -0.5439 -0.1910  0.4832  0.6455 -0.5915  0.3194  0.1307  0.2829
#> 
#> Columns 10 to 18  0.5604 -0.1381 -0.5006 -0.4644  0.1092 -0.2850 -0.0590  0.0288 -0.1488
#>  -0.2401 -0.3703  0.7838  0.1135  0.2279 -0.5429 -0.4211  0.1309 -0.3424
#>  -0.0098 -0.1400 -0.0086  0.5392 -0.4810 -0.2999 -0.2786 -0.1675 -0.5709
#> 
#> Columns 19 to 20 -0.1179  0.0884
#>  -0.5719 -0.3680
#>   0.3750  0.1607
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3249 -0.0855  0.6628  0.5233  0.0015  0.4756  0.5609  0.2259 -0.3769
#>   0.3573  0.0840  0.0678 -0.2342 -0.2440  0.1115  0.3411  0.5605 -0.1234
#>   0.6674  0.3430  0.2986  0.1414  0.0858  0.2287  0.4169  0.2676 -0.2083
#> 
#> Columns 10 to 18 -0.3191 -0.3991 -0.3323 -0.4123  0.3615  0.0612 -0.3198  0.0678  0.0985
#>  -0.4275 -0.2627 -0.1586 -0.2389  0.0162 -0.3899  0.3499 -0.0707 -0.6308
#>  -0.1400  0.0931 -0.5612 -0.1671 -0.0348  0.0084  0.1611 -0.6331 -0.2252
#> 
#> Columns 19 to 20  0.5370  0.2278
#>   0.3162 -0.1500
#>   0.6260  0.1751
#> [ CPUFloatType{2,3,20} ]
#>