RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4821 -0.6100 -0.3400 -0.7237 -0.0776 -0.1100  0.2868  0.5184  0.6926
#>  -0.3221 -0.1370 -0.4227  0.5617  0.5409 -0.7567  0.5028  0.3118  0.0236
#>  -0.8020 -0.4783 -0.0939 -0.3210 -0.2128  0.5195  0.6053  0.9499  0.0587
#> 
#> Columns 10 to 18 -0.3985 -0.4058  0.8313  0.7096  0.1374 -0.4959  0.5791  0.0015  0.2765
#>   0.5824  0.7097  0.6539  0.9275  0.5949 -0.3931 -0.4511 -0.0424  0.4723
#>  -0.8137 -0.4095  0.8371  0.5560  0.3663  0.1920  0.2635  0.1594 -0.4104
#> 
#> Columns 19 to 20 -0.4309 -0.1152
#>  -0.5264  0.6848
#>  -0.7840  0.4784
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4552 -0.3976  0.1687  0.1844  0.4855  0.1390 -0.0974  0.2402  0.1867
#>   0.1098 -0.1424 -0.5047 -0.1229  0.3756  0.0803 -0.0144 -0.5151 -0.0865
#>   0.1367 -0.4972 -0.0563 -0.0494  0.3738 -0.5425  0.1634  0.4113  0.3432
#> 
#> Columns 10 to 18  0.1491  0.0944  0.4418 -0.4852 -0.5285 -0.0309  0.8116  0.3069  0.3997
#>   0.6986 -0.1892  0.1650  0.1098 -0.2338 -0.3006  0.1015 -0.1595  0.3832
#>   0.2954  0.2440  0.1614  0.0810 -0.3053 -0.1041  0.7646 -0.1838  0.2630
#> 
#> Columns 19 to 20 -0.6959  0.2971
#>  -0.5332  0.4169
#>  -0.1735 -0.0342
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3257 -0.7895  0.3832  0.3655  0.0646  0.0452  0.0423 -0.2707  0.0524
#>  -0.5801 -0.5204  0.1390  0.1421  0.0488 -0.2043  0.1739  0.1735  0.2730
#>   0.1010 -0.5530 -0.2297 -0.0298  0.0417  0.0596  0.5282  0.1848  0.3342
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6298 -0.2939 -0.0926  0.3111 -0.2314 -0.0995  0.1698  0.2142 -0.7310
#>   0.8798 -0.2430 -0.6353 -0.1474 -0.1555 -0.1470 -0.3399  0.5045 -0.7948
#>  -0.3540  0.6933  0.1925  0.3900 -0.3638  0.7572 -0.0188  0.0658 -0.5324
#> 
#> Columns 10 to 18  0.2727 -0.2888 -0.3300 -0.2181 -0.0331  0.7334  0.1916  0.0575 -0.0821
#>   0.6047 -0.4475 -0.4670 -0.5789 -0.2771  0.5523 -0.0035 -0.1921  0.5264
#>  -0.1798  0.1383  0.5653 -0.7161 -0.1402 -0.1370 -0.3266  0.5202  0.1303
#> 
#> Columns 19 to 20 -0.1100 -0.1906
#>  -0.7588  0.8904
#>  -0.5070  0.6032
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3424 -0.5071  0.3027  0.3803  0.1893 -0.2078 -0.0012  0.1513  0.3586
#>  -0.0960 -0.3379  0.1105  0.5156  0.0485 -0.4202  0.0290 -0.0550 -0.0326
#>  -0.6647 -0.5010  0.2113  0.5281  0.5130  0.4311 -0.1911 -0.1898 -0.1830
#> 
#> Columns 10 to 18  0.1392 -0.1993  0.0839  0.2454 -0.2760 -0.5128 -0.1288 -0.0231  0.3131
#>  -0.2390  0.3068  0.5730  0.1545 -0.0145 -0.6154 -0.1703 -0.1183 -0.0627
#>  -0.1134  0.2906  0.6606 -0.5131 -0.5254  0.0383  0.0872  0.0622 -0.3408
#> 
#> Columns 19 to 20 -0.4401  0.1680
#>  -0.2415  0.2282
#>   0.0410  0.1134
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>