RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0181 -0.5692 -0.8754 -0.2546 -0.4288  0.0709 -0.3844 -0.1165 -0.0548
#>  -0.1678  0.3730 -0.0524  0.6875 -0.0869  0.8897 -0.8561 -0.4956 -0.6826
#>  -0.2758 -0.0046 -0.6336 -0.3177 -0.3797  0.5790 -0.5738 -0.5007  0.0536
#> 
#> Columns 10 to 18  0.5308  0.0613 -0.2875  0.7263  0.2523 -0.6988 -0.0006 -0.2851  0.4996
#>   0.0912  0.1929 -0.0070  0.7865  0.0356 -0.4621 -0.5526  0.3671  0.1503
#>  -0.3825  0.2925 -0.6957  0.8537  0.7231  0.1398 -0.6575 -0.4917 -0.8003
#> 
#> Columns 19 to 20  0.6657  0.1972
#>   0.4079 -0.4436
#>   0.8865 -0.5479
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1850  0.1468 -0.0995 -0.2290 -0.2148 -0.4311  0.0663 -0.1638  0.2092
#>  -0.1486  0.2404 -0.7318 -0.1849 -0.3248 -0.2389 -0.3045 -0.6260  0.2102
#>   0.0407  0.0140 -0.4792 -0.0940 -0.4390  0.3664 -0.0754  0.4015  0.4906
#> 
#> Columns 10 to 18  0.5739  0.2613  0.6121  0.2367 -0.0496  0.1297 -0.3203 -0.0331  0.0197
#>   0.2894 -0.4018  0.3986  0.5251  0.1183  0.3031 -0.0269  0.1719 -0.3601
#>   0.0652 -0.1641 -0.0966 -0.1865 -0.3138  0.2633 -0.5571 -0.1917 -0.3804
#> 
#> Columns 19 to 20  0.3388 -0.0679
#>   0.4993 -0.0897
#>  -0.1469 -0.1629
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3324  0.5652 -0.0444  0.1884 -0.3467  0.0710  0.3907 -0.0699  0.0266
#>  -0.1297  0.5331 -0.3308  0.0462  0.0123  0.2481 -0.2895 -0.3513 -0.0295
#>   0.2409 -0.4950 -0.3988  0.2634 -0.1938  0.3282 -0.3096 -0.3874  0.5265
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2875  0.2657  0.2765  0.3673  0.4870 -0.7749  0.6518  0.3646  0.0297
#>   0.6051  0.7892  0.0518  0.1642  0.5981 -0.5443 -0.0890  0.2631  0.1539
#>   0.2235  0.8854 -0.4296 -0.7442 -0.6222  0.5220  0.2757 -0.6582  0.4518
#> 
#> Columns 10 to 18  0.1774  0.1139  0.6301 -0.4708  0.0341  0.2505  0.5543 -0.0400 -0.1360
#>   0.3297 -0.1166 -0.2291 -0.1012 -0.0711  0.4607  0.5698 -0.5389 -0.5535
#>  -0.7353  0.1495  0.3503 -0.1863 -0.8815  0.2463  0.8868  0.3931  0.4533
#> 
#> Columns 19 to 20 -0.0518 -0.2851
#>  -0.4908  0.2403
#>   0.1697  0.6493
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5114  0.3014 -0.2378  0.3189 -0.2360  0.0559 -0.2873 -0.4965 -0.2919
#>   0.1018  0.1007  0.1678  0.3597  0.0579 -0.3444  0.2135 -0.5810  0.0285
#>  -0.0400 -0.4075  0.1878 -0.0932 -0.5863  0.4013 -0.0873 -0.1459  0.3657
#> 
#> Columns 10 to 18  0.5307 -0.4702  0.1339  0.4866 -0.1393  0.3147 -0.2661 -0.1913 -0.2796
#>   0.3470 -0.1638  0.4157  0.2932  0.3432  0.4653  0.0156 -0.1776  0.1075
#>   0.2115  0.2463 -0.2565 -0.2803 -0.1768 -0.0083  0.1782 -0.0484  0.3536
#> 
#> Columns 19 to 20  0.4225 -0.4280
#>   0.1572 -0.0263
#>  -0.5086  0.4480
#> [ CPUFloatType{2,3,20} ]
#>