RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1949  0.3105 -0.2114  0.3698 -0.2741  0.4464 -0.0380  0.5889 -0.2792
#>   0.8868 -0.7596 -0.5547 -0.3129  0.5847  0.9166 -0.6634  0.4658 -0.2484
#>  -0.1517  0.9354 -0.5812  0.3492  0.4349  0.8124  0.6206 -0.2997  0.1712
#> 
#> Columns 10 to 18  0.5662  0.4337  0.2844  0.1178 -0.2811  0.0972 -0.5212 -0.7404  0.3652
#>   0.6827 -0.7673 -0.8599 -0.7066 -0.3507 -0.7309  0.1648 -0.4841 -0.0980
#>   0.1164 -0.4253  0.5824  0.3907 -0.4587  0.2986  0.6644 -0.8488 -0.0541
#> 
#> Columns 19 to 20 -0.6343 -0.5949
#>  -0.1827 -0.0005
#>   0.9820  0.8665
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2360  0.3759 -0.3902  0.1846 -0.0204  0.3512 -0.6175 -0.0122  0.2952
#>   0.0675  0.5424 -0.5618  0.0216 -0.0130  0.3450 -0.1505  0.0497  0.4103
#>   0.1160  0.5575 -0.0388  0.4247  0.6413  0.1974 -0.3516  0.4865  0.0129
#> 
#> Columns 10 to 18  0.2906 -0.5986  0.1690 -0.1712 -0.5919 -0.0025  0.2806 -0.2081  0.1576
#>   0.1106 -0.8005 -0.1426  0.3288 -0.6646 -0.0448 -0.4984 -0.1249  0.4785
#>   0.7000 -0.4419  0.6691  0.5838 -0.7619 -0.4482 -0.0593 -0.5039  0.1536
#> 
#> Columns 19 to 20  0.4333  0.5617
#>   0.3915 -0.1447
#>  -0.2981 -0.1422
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2413  0.3615 -0.2748  0.3194 -0.4047 -0.1361 -0.1051 -0.3019  0.1648
#>   0.2224 -0.3683 -0.4496  0.3145  0.4494  0.4145  0.3649  0.3935  0.2390
#>  -0.1907  0.6779 -0.7998  0.2398  0.4415  0.1269 -0.6769  0.3597  0.0267
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1023 -0.0207 -0.0128 -0.7821 -0.0895 -0.2209  0.4075 -0.0901 -0.5408
#>   0.1123 -0.1373  0.4329 -0.1178 -0.1065 -0.3858  0.0726 -0.4024  0.5554
#>   0.3727 -0.3842 -0.0360 -0.4590 -0.0385 -0.6220  0.2421  0.1412 -0.6606
#> 
#> Columns 10 to 18  0.6882  0.0571 -0.4134  0.5766 -0.1583 -0.0359  0.5744 -0.4388 -0.0713
#>   0.3605  0.3047  0.1506 -0.1680  0.1066  0.2692  0.6320 -0.3615 -0.3470
#>  -0.2682  0.8508 -0.4058  0.1239 -0.2826  0.1176 -0.7844 -0.6611 -0.2151
#> 
#> Columns 19 to 20 -0.2871 -0.1034
#>   0.5522  0.3300
#>  -0.6049 -0.5565
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0783  0.1445 -0.4339 -0.1232 -0.2350 -0.4022 -0.1854 -0.3199 -0.0761
#>   0.0743  0.2216 -0.6627  0.2077  0.5166  0.0388 -0.4779  0.1005  0.0758
#>  -0.0246  0.1693 -0.6189 -0.0161 -0.0205 -0.5906 -0.1085  0.2320  0.3779
#> 
#> Columns 10 to 18  0.1954 -0.7680  0.1187 -0.2681 -0.0662  0.1319  0.0634  0.2203  0.5273
#>   0.4735 -0.3144 -0.0997  0.1010 -0.4922  0.1308  0.2484 -0.3006  0.3961
#>   0.0310 -0.2814 -0.0951  0.1952 -0.5040  0.1120 -0.3615  0.0417  0.3587
#> 
#> Columns 19 to 20  0.6045 -0.3693
#>   0.4873 -0.3906
#>   0.5017  0.3064
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>