RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8736 -0.4226 -0.7080  0.7255  0.7738  0.0864  0.7039  0.8822  0.4835
#>   0.7655  0.3438 -0.6332  0.5905 -0.6712 -0.6330  0.2825  0.3185  0.4717
#>  -0.7007 -0.5317 -0.2756 -0.1241  0.8373  0.1173  0.1327 -0.1967  0.2818
#> 
#> Columns 10 to 18  0.4815  0.2178 -0.7289  0.5471  0.4227 -0.0644  0.1228  0.3583 -0.2732
#>  -0.8839 -0.2536  0.5336 -0.3706 -0.2365  0.5693 -0.3090 -0.2219  0.6944
#>   0.5028 -0.0807 -0.0391  0.9618  0.5620 -0.0493  0.7917  0.0287  0.1470
#> 
#> Columns 19 to 20 -0.0158  0.4790
#>   0.0332  0.7632
#>  -0.2025  0.6769
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0915 -0.2670  0.0835 -0.3262 -0.0719 -0.0224  0.3641 -0.1948 -0.2335
#>   0.2435 -0.4706  0.5959 -0.8347 -0.4236 -0.5506  0.3506  0.2814 -0.2742
#>  -0.0047 -0.6459  0.1810 -0.2659  0.3536 -0.0306  0.1378 -0.0763  0.0461
#> 
#> Columns 10 to 18 -0.6751  0.5244 -0.3087  0.4747  0.1156  0.8022 -0.6331 -0.2608  0.4634
#>  -0.6957  0.3724  0.1714  0.6139 -0.0886  0.4988  0.5394  0.2625  0.5843
#>   0.1602 -0.0065 -0.0257  0.2239 -0.2925  0.6491  0.4215 -0.6996 -0.1393
#> 
#> Columns 19 to 20  0.4987  0.1838
#>   0.1219  0.7248
#>   0.7402  0.1083
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3905 -0.1705 -0.3002 -0.2139  0.1035 -0.4437  0.1308 -0.3979 -0.0777
#>  -0.0355 -0.7501 -0.5437  0.0431 -0.1556 -0.0600  0.2088 -0.1293 -0.1440
#>   0.0908 -0.1418  0.3656 -0.1727 -0.0378  0.3259 -0.0344 -0.4437 -0.0559
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4284  0.2141 -0.3260  0.3435  0.4746  0.4968  0.7866 -0.2358  0.4163
#>  -0.2986  0.4753 -0.6677  0.4115 -0.4390  0.2396 -0.1793  0.7227 -0.5392
#>   0.6473  0.1142 -0.4045 -0.7730  0.2230  0.2925  0.3936 -0.6007 -0.0983
#> 
#> Columns 10 to 18 -0.4141 -0.1503  0.0407 -0.7217 -0.2600  0.2511  0.4265 -0.4479  0.5118
#>   0.5608  0.3522 -0.7004  0.5592  0.4789 -0.8989 -0.2563 -0.7580 -0.2304
#>  -0.4862 -0.7859 -0.6881 -0.4733  0.4520  0.2428 -0.1078 -0.6969  0.7727
#> 
#> Columns 19 to 20 -0.6410  0.6562
#>  -0.2762  0.5168
#>  -0.9054  0.7535
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1055 -0.4357  0.0904 -0.2750 -0.2321 -0.2360  0.3445 -0.0126  0.0455
#>   0.1280 -0.1395 -0.6998 -0.4354  0.1672 -0.5906  0.1276  0.1600  0.2586
#>   0.2735 -0.6576  0.4342 -0.5722 -0.4381 -0.3283  0.4697  0.2475  0.1588
#> 
#> Columns 10 to 18 -0.6662  0.1732  0.1566  0.1699 -0.2662  0.5793  0.0930  0.2887  0.1850
#>   0.2832  0.0241  0.4121  0.6626 -0.3621  0.1290  0.1529 -0.2120  0.2614
#>  -0.5214  0.4910 -0.1858  0.4187  0.3549  0.7321  0.5316 -0.1749  0.1605
#> 
#> Columns 19 to 20  0.4875  0.5543
#>  -0.5359  0.1211
#>   0.4939  0.4463
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>