RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5384 -0.0086  0.4132  0.7894 -0.2911 -0.2327  0.5043  0.9705  0.0779
#>  -0.8159  0.1750  0.4589  0.1358  0.8005  0.7550 -0.1765  0.7910 -0.4179
#>   0.5694  0.7102  0.1369 -0.4912 -0.6294  0.7157  0.0555  0.5924  0.2232
#> 
#> Columns 10 to 18 -0.6611 -0.1964  0.1782  0.4634  0.4095  0.8671  0.4468 -0.6477 -0.4736
#>   0.7546 -0.7957  0.0803  0.2504  0.3326 -0.5431  0.1627 -0.9004 -0.0239
#>  -0.2696 -0.0520 -0.7606  0.4826  0.4359 -0.4410 -0.2312 -0.8826 -0.4443
#> 
#> Columns 19 to 20  0.2572 -0.3002
#>  -0.7695 -0.5632
#>   0.2142  0.6661
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4931  0.3113  0.3467  0.7062  0.2206  0.3227 -0.2087  0.2558 -0.2520
#>  -0.1172  0.5977 -0.3157  0.4184 -0.2123 -0.2808  0.5199  0.4341 -0.4657
#>   0.2774  0.2258  0.3923  0.4909 -0.2337  0.1151 -0.3391  0.1877 -0.0238
#> 
#> Columns 10 to 18 -0.5138 -0.5219 -0.1264 -0.4554  0.3335 -0.1645 -0.0732  0.1254  0.5895
#>  -0.2645 -0.5247 -0.3323  0.2490 -0.2006 -0.6585  0.0194 -0.4840 -0.0232
#>  -0.3427 -0.4951 -0.0749  0.7168  0.1929 -0.3091 -0.3906 -0.5225  0.4202
#> 
#> Columns 19 to 20  0.7082  0.0991
#>   0.5257  0.2947
#>   0.5980 -0.6327
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2220  0.0793  0.2160  0.0816  0.1473  0.5734  0.2961  0.4643 -0.4299
#>   0.1802  0.0770  0.2429  0.3561 -0.1002  0.1655  0.6633 -0.3635 -0.3214
#>  -0.0189  0.0032  0.4848  0.5468  0.0948  0.0824  0.2754  0.0057 -0.2659
#> 
#> Columns 10 to 18 -0.0713 -0.2454 -0.6174 -0.4375  0.1837 -0.2171 -0.1047 -0.5617 -0.1692
#>  -0.3268 -0.2149  0.2678 -0.4254 -0.1617 -0.3671 -0.5997  0.0470  0.4356
#>   0.0218 -0.2271 -0.3380 -0.3554  0.0054 -0.1997 -0.4153  0.0424  0.2010
#> 
#> Columns 19 to 20  0.3479 -0.1113
#>   0.6244 -0.1818
#>   0.7301 -0.0814
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.5690  0.2778  0.2658  0.1610 -0.2907  0.3468  0.5286  0.3174 -0.4397
#>   0.4338 -0.0522  0.4257  0.2733 -0.2433  0.4885  0.5766  0.5577 -0.3694
#>   0.3086  0.5204  0.5531  0.3430 -0.2798 -0.0198  0.3925  0.5639 -0.1559
#> 
#> Columns 10 to 18 -0.5729 -0.4778 -0.4617  0.2818  0.2600 -0.3312  0.1489 -0.5189 -0.2309
#>  -0.0460 -0.0016 -0.3612 -0.6683  0.0838 -0.1171 -0.2082 -0.0370  0.0184
#>  -0.3869 -0.1927 -0.0902 -0.4140  0.1960 -0.2227 -0.1217 -0.0136  0.0580
#> 
#> Columns 19 to 20  0.4329 -0.0606
#>   0.6311  0.5021
#>   0.6620  0.2144
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.7115  0.1348  0.1341  0.3513 -0.1259  0.5617  0.3636  0.1148 -0.0116
#>   0.5548  0.2246  0.2024  0.1180  0.0406  0.3200  0.4145  0.3332 -0.4330
#>   0.3699  0.0609  0.3578  0.2799 -0.1087  0.7306  0.3920  0.5950 -0.2741
#> 
#> Columns 10 to 18 -0.2441 -0.3148 -0.1248  0.0533 -0.0740 -0.1711 -0.2926 -0.5083 -0.3279
#>  -0.4468 -0.5880 -0.5107 -0.1498  0.0278 -0.2231  0.1019 -0.4530 -0.3381
#>  -0.2598 -0.3447 -0.4123 -0.2063  0.4055 -0.1573  0.0545 -0.5282  0.0019
#> 
#> Columns 19 to 20  0.5109  0.0187
#>   0.4722 -0.2600
#>   0.5568  0.0364
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2628 -0.0709  0.3080  0.4029  0.3731 -0.2224  0.5249  0.2844 -0.3298
#>  -0.5723 -0.2058 -0.4853  0.5010 -0.3220  0.5896 -0.3468 -0.1876 -0.2374
#>   0.3369 -0.1605 -0.4410  0.4290  0.2699  0.6264 -0.0563 -0.4863  0.0889
#> 
#> Columns 10 to 18 -0.0889 -0.1904 -0.5520  0.4510 -0.8052 -0.4429  0.0354  0.5631 -0.6898
#>   0.3550 -0.3090  0.1362  0.1070  0.0856 -0.2750 -0.1404  0.3956 -0.1952
#>   0.1467 -0.3468 -0.2887  0.4379 -0.2899 -0.5418 -0.5953 -0.6131  0.2308
#> 
#> Columns 19 to 20  0.1965 -0.5289
#>  -0.7152  0.4279
#>  -0.4694  0.2513
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.7115  0.1348  0.1341  0.3513 -0.1259  0.5617  0.3636  0.1148 -0.0116
#>   0.5548  0.2246  0.2024  0.1180  0.0406  0.3200  0.4145  0.3332 -0.4330
#>   0.3699  0.0609  0.3578  0.2799 -0.1087  0.7306  0.3920  0.5950 -0.2741
#> 
#> Columns 10 to 18 -0.2441 -0.3148 -0.1248  0.0533 -0.0740 -0.1711 -0.2926 -0.5083 -0.3279
#>  -0.4468 -0.5880 -0.5107 -0.1498  0.0278 -0.2231  0.1019 -0.4530 -0.3381
#>  -0.2598 -0.3447 -0.4123 -0.2063  0.4055 -0.1573  0.0545 -0.5282  0.0019
#> 
#> Columns 19 to 20  0.5109  0.0187
#>   0.4722 -0.2600
#>   0.5568  0.0364
#> [ CPUFloatType{2,3,20} ]
#>