RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2087  0.3296  0.9009 -0.2149 -0.3668  0.0642 -0.4213 -0.1581  0.6754
#>  -0.6599 -0.8204  0.2897  0.2747  0.4532 -0.0610 -0.1149 -0.3829  0.8637
#>  -0.0866  0.2901 -0.7065  0.3512  0.5624  0.1269  0.6177  0.3417  0.2726
#> 
#> Columns 10 to 18  0.0269 -0.5368  0.1488  0.4592 -0.1827  0.8223  0.6940  0.5358 -0.4584
#>  -0.1391 -0.5915  0.2895 -0.3539  0.7993 -0.1926  0.8694 -0.3288 -0.0488
#>   0.5836 -0.1225  0.3583 -0.1415 -0.0811  0.3239 -0.6421 -0.2387  0.0620
#> 
#> Columns 19 to 20  0.6114 -0.0295
#>  -0.4726 -0.1243
#>   0.4312 -0.3362
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5012  0.3807  0.0546 -0.1460 -0.0256 -0.0428 -0.6552  0.0986  0.3746
#>   0.2992  0.3445  0.3334 -0.0473 -0.2806 -0.3500 -0.1455 -0.2229  0.1719
#>   0.1145 -0.1577  0.3959 -0.1765 -0.1876 -0.1369 -0.3930 -0.2995  0.5566
#> 
#> Columns 10 to 18 -0.7353 -0.0640  0.4764 -0.7127 -0.5027 -0.1101  0.1019  0.3243  0.7810
#>  -0.3021  0.3380 -0.2341 -0.1126  0.2479  0.0229  0.3797  0.4466  0.2586
#>  -0.2991  0.4632  0.4914 -0.1154  0.3740  0.3325  0.5687  0.0101 -0.1279
#> 
#> Columns 19 to 20  0.3518 -0.3372
#>   0.4254 -0.5673
#>  -0.1992  0.3379
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4897  0.3032 -0.1502  0.0293 -0.2352 -0.4432 -0.4584  0.1089  0.1615
#>  -0.2591 -0.1478  0.3510  0.2804  0.2806 -0.1625 -0.2856 -0.1560  0.5740
#>  -0.2279 -0.3932  0.5478  0.0966 -0.4988 -0.1243 -0.5847  0.3997  0.3228
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1649  0.1817 -0.4737  0.4403 -0.1309  0.1013 -0.5007 -0.6501 -0.5121
#>   0.1937 -0.0689  0.1574 -0.7783 -0.6378 -0.6115  0.5504 -0.3909  0.4035
#>   0.8240 -0.1511 -0.4183 -0.7483 -0.6783  0.3036  0.2913 -0.5029  0.6126
#> 
#> Columns 10 to 18 -0.1873  0.2606  0.4299  0.3824  0.1301  0.0270 -0.3715  0.0729  0.0141
#>   0.5973  0.3794 -0.6289  0.0872 -0.2280  0.3638  0.3959  0.5230  0.6889
#>   0.2742  0.3600 -0.6836 -0.5886 -0.6432  0.0995  0.1881  0.8752  0.4850
#> 
#> Columns 19 to 20 -0.0818  0.3611
#>   0.5189 -0.3115
#>  -0.4382  0.4888
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2540  0.2941 -0.0971  0.4577  0.3760 -0.2937 -0.4622  0.3124  0.0291
#>  -0.1475  0.4782  0.0361 -0.0814  0.1687 -0.4585  0.0076  0.2019 -0.1255
#>  -0.1967  0.4055  0.0450 -0.0931 -0.1405 -0.5634 -0.2521  0.4038  0.0537
#> 
#> Columns 10 to 18  0.1287  0.0255 -0.0731 -0.4122 -0.5445 -0.1614  0.3162  0.2608  0.2024
#>  -0.3059 -0.0316 -0.0848 -0.4186 -0.4603 -0.1602  0.5529  0.3293  0.2791
#>  -0.6755 -0.0355  0.3071 -0.7138 -0.4881 -0.6104  0.3803  0.3905  0.3866
#> 
#> Columns 19 to 20 -0.0161 -0.0912
#>   0.3032 -0.1480
#>   0.2503 -0.1605
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>