RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1445  0.6113  0.4694  0.0845  0.5150  0.3497 -0.4230  0.3018  0.2690
#>  -0.6665 -0.4788  0.7416 -0.4651  0.2176  0.0129 -0.8569  0.8390 -0.4333
#>  -0.6724 -0.5619  0.2676 -0.2800 -0.7027 -0.0831  0.1032  0.5173  0.2310
#> 
#> Columns 10 to 18 -0.3592 -0.0965  0.3902  0.4954  0.3372  0.2899 -0.6289 -0.5761 -0.2062
#>  -0.1950  0.1570  0.6607  0.2446 -0.0209  0.7112 -0.4036 -0.0025  0.3824
#>  -0.3806 -0.2174 -0.3788 -0.8215 -0.3513 -0.4853  0.3722  0.6495 -0.4381
#> 
#> Columns 19 to 20  0.3720  0.2356
#>  -0.2598 -0.4246
#>   0.2260  0.4096
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2571  0.2082 -0.6844 -0.2930 -0.3006  0.4596  0.0485 -0.0764  0.5422
#>  -0.2787 -0.1393 -0.6987 -0.1077 -0.3583  0.6418  0.3003 -0.6973  0.5660
#>  -0.3665  0.0781  0.0343 -0.0806 -0.4358  0.0170  0.1723 -0.2282 -0.0192
#> 
#> Columns 10 to 18 -0.3145 -0.1518 -0.3878 -0.0143 -0.1138  0.1454 -0.6136 -0.2731  0.3879
#>  -0.0782  0.0468  0.0709 -0.2532 -0.4862  0.4252 -0.0905 -0.4142  0.3886
#>  -0.1555  0.2133 -0.1539 -0.3756  0.0521  0.4634 -0.0141 -0.1222  0.0620
#> 
#> Columns 19 to 20 -0.2731  0.7046
#>  -0.4127  0.5861
#>  -0.1760 -0.0556
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1644 -0.3562 -0.0434 -0.2308 -0.4863  0.1408 -0.2602  0.4526 -0.3235
#>  -0.5187  0.0463 -0.0106 -0.0195 -0.0063  0.0650 -0.5282  0.3761 -0.3242
#>  -0.0502  0.0759 -0.3104 -0.1519 -0.0752  0.5856 -0.1031  0.2540  0.4081
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7089 -0.5248  0.4306 -0.1899  0.0948 -0.6178 -0.1832  0.1696  0.5136
#>  -0.3974 -0.2317  0.0444 -0.7450 -0.1088 -0.6027  0.4218  0.4376  0.0344
#>  -0.5468 -0.3200  0.1595  0.0727  0.0897 -0.0911  0.1383  0.0332  0.1861
#> 
#> Columns 10 to 18  0.5745  0.4481 -0.6439  0.8239  0.2673 -0.5047  0.4967  0.7713 -0.1473
#>  -0.1216  0.1159  0.2133  0.0331  0.1705 -0.0545 -0.0646 -0.0200 -0.1575
#>   0.0944  0.4486 -0.4456  0.4124  0.4357 -0.5332  0.0799  0.2904 -0.4527
#> 
#> Columns 19 to 20 -0.5287  0.1868
#>   0.7074 -0.8259
#>   0.2765  0.2029
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2639 -0.1635 -0.2935  0.0852 -0.1818  0.2958  0.1065 -0.1840  0.3803
#>  -0.5199 -0.1725 -0.1546  0.0907 -0.3694  0.1997 -0.1351 -0.3485  0.5209
#>  -0.2432 -0.2254 -0.0403  0.1729 -0.0493  0.1890 -0.3174  0.0701  0.2193
#> 
#> Columns 10 to 18 -0.6575  0.0135 -0.5094 -0.1079 -0.6019  0.6999 -0.5321 -0.4919  0.1655
#>  -0.5297  0.4444 -0.2446  0.2821 -0.2671  0.2872  0.2356 -0.6738  0.3792
#>  -0.3995  0.4501 -0.1657  0.1891 -0.3950  0.4768 -0.2894 -0.6085  0.4159
#> 
#> Columns 19 to 20 -0.4968  0.1579
#>  -0.1703 -0.1340
#>  -0.6169 -0.1406
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>