RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5110  0.7251  0.1925 -0.1915 -0.1258 -0.1116  0.2345 -0.0563 -0.5666
#>  -0.3153  0.4233 -0.0969  0.7350  0.1543  0.7343  0.1512 -0.6063  0.0523
#>   0.0501  0.7621  0.4878  0.2092 -0.2382 -0.2126 -0.4977  0.7792  0.0740
#> 
#> Columns 10 to 18  0.3515  0.6105  0.6410 -0.1715  0.2341 -0.6651  0.5334 -0.7813  0.5167
#>  -0.7451  0.2379 -0.5980  0.5750  0.6114 -0.5949 -0.5500 -0.1772  0.6084
#>   0.6089 -0.2510  0.8444 -0.8739  0.0917  0.0852  0.2893 -0.8663 -0.0221
#> 
#> Columns 19 to 20 -0.1549  0.8802
#>   0.2424  0.1907
#>  -0.3006  0.5332
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6769  0.0121 -0.2197  0.2201 -0.1302  0.3356 -0.4979  0.2611 -0.3006
#>  -0.5767  0.1511 -0.2091  0.8802 -0.2458 -0.2417 -0.3867 -0.1282  0.3048
#>  -0.3636 -0.2717  0.6298  0.1539  0.0638  0.4260  0.0421  0.1177 -0.4346
#> 
#> Columns 10 to 18  0.4600  0.2597  0.0897  0.6208 -0.2982 -0.1627  0.0302  0.5120 -0.4656
#>   0.3190  0.1022  0.4719 -0.0773  0.4884 -0.2262  0.5972 -0.2329 -0.3514
#>   0.2554  0.2500 -0.3897  0.6608 -0.3360 -0.5237 -0.5604  0.7181  0.0352
#> 
#> Columns 19 to 20  0.6008 -0.0202
#>   0.2536 -0.1182
#>   0.2636  0.1070
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0993 -0.4203  0.3207  0.2915  0.1537 -0.0897 -0.0458 -0.0614 -0.0052
#>  -0.0806  0.0466  0.3368  0.4389 -0.1656  0.0492  0.0091  0.0038 -0.0496
#>  -0.0952 -0.4069 -0.2093  0.0789  0.6171  0.0240 -0.4796  0.1295 -0.1224
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1747  0.2076  0.1724 -0.5521  0.0384  0.5132 -0.4289  0.3718 -0.2686
#>  -0.0924 -0.6070 -0.4013 -0.6277 -0.1880  0.2349 -0.0571 -0.0297  0.0047
#>   0.4765  0.5464  0.0708 -0.2958  0.1517  0.7855 -0.5703  0.4092  0.6414
#> 
#> Columns 10 to 18 -0.3119 -0.0749  0.8698  0.2867  0.6463  0.8058 -0.4817  0.4127  0.5172
#>  -0.0440  0.2013  0.8166 -0.3426  0.4029  0.7579 -0.0518 -0.1241  0.3568
#>  -0.3532  0.0460 -0.5210  0.0119 -0.3310 -0.6099 -0.4732 -0.8993 -0.1437
#> 
#> Columns 19 to 20  0.7689  0.7233
#>   0.3599  0.8949
#>   0.2883 -0.8890
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4947 -0.3956 -0.0112 -0.1890  0.6011  0.2002 -0.0582 -0.2527 -0.7119
#>   0.4579 -0.2603  0.2880 -0.1606  0.6088  0.4241 -0.2330  0.0106 -0.4572
#>  -0.3940 -0.4111  0.1955 -0.2756  0.1649  0.1070 -0.1351  0.3780 -0.4901
#> 
#> Columns 10 to 18 -0.2807  0.6628  0.0160 -0.1798 -0.5277  0.4103 -0.3923  0.3103  0.0604
#>   0.2104  0.2546 -0.1462  0.0387 -0.3391  0.1961 -0.6930  0.5900 -0.4164
#>   0.3405  0.4223  0.1920 -0.1070  0.2049 -0.1471 -0.0983 -0.2370  0.0127
#> 
#> Columns 19 to 20  0.5908  0.5512
#>   0.2608  0.7701
#>   0.1207  0.4063
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>