RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7619 -0.4224 -0.5983 -0.8200 -0.0443 -0.0669  0.0770  0.7472 -0.5153
#>   0.3698  0.4834  0.0660  0.6726  0.9386 -0.7598  0.6610 -0.7670 -0.0558
#>   0.2346 -0.4024  0.7150 -0.8977  0.2888  0.0669 -0.8422  0.6575  0.5556
#> 
#> Columns 10 to 18  0.4066 -0.3584 -0.2494 -0.3810 -0.5539  0.6630  0.5609  0.7628 -0.3353
#>  -0.5494  0.2114 -0.6756 -0.6861  0.7495 -0.8224  0.3321 -0.9269 -0.1729
#>   0.1123 -0.9675 -0.1507  0.9156 -0.6518  0.8208 -0.6353  0.2183 -0.7333
#> 
#> Columns 19 to 20  0.6432  0.6672
#>  -0.0711 -0.7551
#>   0.2103  0.3207
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3773 -0.6054 -0.4061 -0.1307 -0.3763  0.3449  0.0891 -0.1269 -0.4405
#>  -0.2121  0.1756  0.3686 -0.0322 -0.2739 -0.7595  0.7130 -0.7937 -0.2784
#>  -0.1162 -0.0103 -0.3081 -0.2604 -0.4537 -0.0477 -0.7201  0.0048  0.0395
#> 
#> Columns 10 to 18  0.7534 -0.5455  0.3486 -0.2372 -0.4490 -0.7375  0.1292  0.2395  0.1761
#>   0.0771  0.2683 -0.0614  0.1039 -0.1768  0.2340 -0.5214 -0.3652 -0.2253
#>   0.4887 -0.5894  0.3828  0.0267 -0.1766 -0.2907 -0.5023  0.0900  0.4015
#> 
#> Columns 19 to 20  0.0537  0.7009
#>   0.0096 -0.1578
#>  -0.5783  0.3634
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4659 -0.6179 -0.6561 -0.4656  0.6263 -0.1009  0.2554  0.7751  0.2446
#>  -0.3917 -0.2553 -0.5600  0.2533  0.0902 -0.0023  0.4208 -0.4753 -0.4557
#>  -0.6763 -0.3533 -0.1403  0.0133  0.5049  0.2808  0.4056 -0.0833  0.2917
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1421  0.7102 -0.6171 -0.7327 -0.3641  0.2509 -0.0230  0.5039 -0.4763
#>  -0.4964 -0.4447 -0.3705 -0.6603 -0.3348 -0.4526 -0.2344  0.7574 -0.4923
#>  -0.5330  0.1808  0.0659 -0.0670 -0.0590 -0.0242 -0.4422  0.5996 -0.3011
#> 
#> Columns 10 to 18  0.3526  0.6221  0.8649  0.4794  0.4610  0.1939  0.1789  0.5658  0.3990
#>   0.1110  0.3530  0.3904  0.5185 -0.5560  0.5069  0.2858  0.2315  0.1666
#>   0.2565 -0.5705 -0.6776  0.4357  0.1307  0.7147  0.4455 -0.3609  0.6651
#> 
#> Columns 19 to 20  0.1193  0.3727
#>  -0.2571  0.5459
#>   0.1935 -0.2364
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7141 -0.4215 -0.3390  0.0479  0.4639 -0.2194  0.1683 -0.0960  0.3378
#>  -0.4633 -0.3366  0.1074  0.0444  0.1609 -0.2913  0.0736 -0.2576  0.0258
#>  -0.5851 -0.5290 -0.1851  0.0330  0.4569 -0.2574  0.5863  0.0242 -0.2861
#> 
#> Columns 10 to 18  0.3198 -0.5388  0.1682 -0.1808  0.3057 -0.0184  0.0367 -0.2287  0.0156
#>   0.7633 -0.4508  0.3627 -0.1676 -0.5387  0.0190 -0.3058 -0.2021 -0.0490
#>   0.5558 -0.0864  0.1432 -0.4198  0.0093 -0.0879  0.0037 -0.3716 -0.1617
#> 
#> Columns 19 to 20 -0.0459  0.2158
#>  -0.4133  0.0316
#>  -0.0896 -0.3988
#> [ CPUFloatType{2,3,20} ]
#>