RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3315 -0.6987 -0.6774  0.5253  0.5092  0.4813  0.0730 -0.3077 -0.2539
#>   0.1600 -0.8646 -0.4161 -0.0581 -0.4817 -0.1960 -0.3330  0.5277 -0.6176
#>   0.1737  0.5916 -0.5102  0.0731 -0.6243 -0.2181 -0.3215  0.5252  0.4021
#> 
#> Columns 10 to 18 -0.6774  0.7754  0.2809  0.0984 -0.5149  0.1839  0.6663 -0.8664 -0.5341
#>  -0.7359  0.0440  0.2094 -0.7966 -0.3654  0.1975  0.5787 -0.3616  0.4881
#>  -0.1447  0.6158  0.1967  0.1938  0.6865 -0.3600 -0.4256  0.6232  0.6158
#> 
#> Columns 19 to 20  0.2247  0.6447
#>   0.8447  0.6923
#>  -0.6456 -0.7004
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2725 -0.7639  0.0440  0.1179  0.0903  0.0756 -0.0771  0.2871 -0.2878
#>  -0.1720  0.3997 -0.2881 -0.1333 -0.5878 -0.8045 -0.3040  0.6049 -0.2990
#>   0.5155 -0.6008 -0.5599  0.1435  0.1500  0.5893 -0.4323 -0.1711  0.3328
#> 
#> Columns 10 to 18 -0.8167 -0.7899  0.6455  0.0632  0.1763 -0.0817  0.3182  0.6479  0.7281
#>  -0.6901  0.2741 -0.0998  0.5940 -0.0530 -0.6075  0.5452 -0.1431  0.2809
#>   0.3329 -0.1608  0.7046 -0.1766  0.2877  0.7626 -0.0798  0.0342 -0.0039
#> 
#> Columns 19 to 20  0.1209  0.1944
#>   0.0589  0.4095
#>  -0.1072 -0.1367
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3196  0.1059  0.0645 -0.2825 -0.7331 -0.5306 -0.6002  0.4811  0.5042
#>  -0.2133  0.3358  0.0540 -0.2224 -0.6802  0.0577 -0.3608  0.3782  0.4694
#>  -0.1488 -0.7459 -0.4176  0.2686 -0.0133 -0.4971 -0.4455  0.2444 -0.5396
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7986 -0.0161  0.8787 -0.0644  0.5976 -0.6920  0.1729 -0.2636 -0.2609
#>  -0.1993 -0.6431  0.4046 -0.2686 -0.4432  0.1853  0.5311 -0.4468 -0.5128
#>  -0.0264  0.3818  0.7990 -0.0012  0.6253  0.3807 -0.3676 -0.7588 -0.7249
#> 
#> Columns 10 to 18 -0.6797 -0.1133 -0.0789  0.3197 -0.4876 -0.1117 -0.4899  0.7042 -0.2124
#>   0.0365  0.4266 -0.1952  0.1985 -0.4057 -0.5478 -0.5449  0.4998  0.7564
#>  -0.0840 -0.0320 -0.3379 -0.1144 -0.4166  0.4811  0.3598  0.4708 -0.7866
#> 
#> Columns 19 to 20  0.1375 -0.3902
#>   0.2576  0.0778
#>  -0.3515  0.1143
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0600 -0.6191  0.1733 -0.2682 -0.7242 -0.2619 -0.3816  0.4566  0.1098
#>   0.1373 -0.5305 -0.5436 -0.1540 -0.0376 -0.0569  0.1337  0.1263 -0.2822
#>   0.1211  0.4955  0.3110  0.2594 -0.6473 -0.3353 -0.5214  0.0494  0.2482
#> 
#> Columns 10 to 18 -0.5504  0.1558  0.2362  0.3359  0.1076 -0.0565  0.1200  0.1432  0.4557
#>  -0.4560 -0.0613  0.1198  0.1581 -0.1618 -0.5179 -0.2029 -0.0570  0.4068
#>  -0.3312  0.5266  0.4116 -0.1324  0.0244 -0.0066  0.0859 -0.1248 -0.0704
#> 
#> Columns 19 to 20  0.4305 -0.3702
#>   0.1476  0.1252
#>   0.5490 -0.2493
#> [ CPUFloatType{2,3,20} ]
#>