RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1027  0.2159 -0.1798  0.1202  0.0790  0.1669  0.5216  0.1528  0.4153
#>  -0.6086 -0.4618 -0.5413 -0.8570 -0.5210 -0.6651 -0.4975 -0.9284 -0.4148
#>  -0.1929  0.2590 -0.3710 -0.5440 -0.5541  0.0416 -0.1964 -0.4800 -0.1821
#> 
#> Columns 10 to 18 -0.1745  0.0685  0.7175  0.4708 -0.2865 -0.5618  0.7283  0.2584  0.6704
#>   0.5900  0.6515 -0.7425  0.1132 -0.0389 -0.6352  0.0274 -0.4347  0.7508
#>   0.2167 -0.0495 -0.3856 -0.2759 -0.4021 -0.5025  0.6043  0.0609  0.6867
#> 
#> Columns 19 to 20 -0.7142  0.0316
#>   0.0942 -0.5585
#>  -0.4590 -0.2531
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5437 -0.4099 -0.1815 -0.7717  0.3476 -0.1425 -0.1426 -0.6761 -0.4688
#>   0.2304 -0.1071 -0.8027 -0.3110  0.4058 -0.0727  0.5157 -0.5696  0.0837
#>   0.4633 -0.0376 -0.2399 -0.5829 -0.1853  0.2441 -0.5034  0.2995 -0.2855
#> 
#> Columns 10 to 18  0.0714 -0.2589 -0.3563 -0.2472  0.0985 -0.1826  0.4020 -0.3386  0.0334
#>  -0.2757  0.2509 -0.2618 -0.2915  0.6401 -0.3300  0.1371  0.1012 -0.0889
#>   0.0298 -0.7362 -0.5129  0.0626  0.5181 -0.3164 -0.2284 -0.2004  0.0580
#> 
#> Columns 19 to 20 -0.6441 -0.3723
#>  -0.4230 -0.3413
#>  -0.4760 -0.3493
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5255 -0.0996 -0.0184 -0.4248  0.3899 -0.3126  0.4602 -0.5596 -0.2121
#>  -0.2054 -0.1765 -0.5085 -0.5677  0.1273 -0.2648  0.2759 -0.3384 -0.4172
#>  -0.0636 -0.2124  0.1382 -0.3290 -0.4837 -0.1527 -0.2439 -0.3172  0.5649
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0265  0.1294  0.3463 -0.0294  0.6651  0.4209 -0.2199 -0.6239  0.1176
#>  -0.0525 -0.6025  0.1740  0.4234  0.2219 -0.4354 -0.1186  0.4218  0.1884
#>   0.7340 -0.2496  0.1162  0.0001 -0.0136  0.1217  0.3715  0.4663 -0.2840
#> 
#> Columns 10 to 18 -0.7351 -0.3317  0.5798 -0.0497 -0.5309 -0.0770  0.0995 -0.2402  0.0369
#>   0.1231 -0.0263  0.5979  0.7817  0.6578 -0.3376  0.4788 -0.1149  0.4919
#>   0.5640 -0.2596 -0.1917  0.0518  0.1036 -0.0679 -0.7378 -0.6450  0.0507
#> 
#> Columns 19 to 20 -0.3100  0.0048
#>   0.8772 -0.5956
#>   0.4807  0.5817
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0146  0.3999 -0.5352 -0.4986  0.0779 -0.0338  0.5088 -0.5176  0.2040
#>  -0.6636 -0.4000 -0.4058 -0.1915  0.1154  0.0425 -0.2994 -0.5991 -0.2510
#>  -0.5324 -0.1403 -0.3269  0.0490  0.1510 -0.3053 -0.0072 -0.4058 -0.3862
#> 
#> Columns 10 to 18  0.0260  0.3779 -0.0342 -0.3544  0.4156  0.0412  0.0984 -0.4995  0.3958
#>   0.1273  0.1398  0.4568 -0.2655  0.1369 -0.2174  0.2763 -0.4122 -0.0114
#>   0.0662  0.0692  0.1299  0.3523  0.2329 -0.6935 -0.0738 -0.4049  0.1007
#> 
#> Columns 19 to 20 -0.1923 -0.2515
#>  -0.5461 -0.4995
#>  -0.4880 -0.3317
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>