RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5260  0.5172 -0.3798  0.4507 -0.2575 -0.8200  0.9419  0.6077  0.3997
#>  -0.7340 -0.3002 -0.2736 -0.1721  0.7238  0.1248  0.6571 -0.1777 -0.1773
#>   0.7737 -0.5271  0.0513  0.7981  0.8359 -0.2180 -0.0103 -0.1613 -0.6502
#> 
#> Columns 10 to 18  0.5856  0.6845  0.2369  0.4005 -0.8202  0.1454  0.7671  0.1845 -0.6393
#>  -0.4085 -0.5530  0.0853 -0.4202 -0.6902 -0.1991  0.7822  0.0261 -0.6866
#>  -0.6591 -0.8076  0.4490  0.5206 -0.5169 -0.3687 -0.1624 -0.5166 -0.0047
#> 
#> Columns 19 to 20  0.3123 -0.8768
#>  -0.2894 -0.2942
#>   0.1195 -0.5429
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1127  0.5132  0.5255  0.1012  0.0819  0.2426  0.3306  0.7646  0.0247
#>  -0.0045 -0.0629  0.2958 -0.2025  0.3163  0.0901  0.6121  0.2276 -0.5293
#>   0.1434 -0.6443  0.5219 -0.0667  0.3609  0.0691  0.6917 -0.1681 -0.1555
#> 
#> Columns 10 to 18 -0.5707  0.1076 -0.1470 -0.2075  0.7204  0.3835  0.7311 -0.3141  0.1039
#>  -0.3835 -0.2159 -0.5515 -0.3428 -0.2654 -0.0036  0.6854 -0.2954 -0.6118
#>  -0.5395 -0.2123  0.2096 -0.5371 -0.1409 -0.1357 -0.2198  0.2814 -0.5479
#> 
#> Columns 19 to 20  0.1242  0.4045
#>   0.0908 -0.1353
#>   0.2262 -0.1084
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1384  0.1820  0.2591  0.4080  0.4190  0.6195 -0.2785  0.4426  0.1949
#>   0.4303 -0.1513  0.2738  0.1268  0.1519  0.4365  0.3483  0.2515 -0.4265
#>   0.6379  0.0953  0.3848  0.4201 -0.1511 -0.0826  0.6402  0.5419 -0.1717
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5804 -0.3068 -0.2513  0.7593  0.4361 -0.5476  0.6396  0.2986 -0.0405
#>  -0.4287  0.0192  0.3713  0.4419  0.5151 -0.1169 -0.6613  0.3661  0.1401
#>   0.4096 -0.6897 -0.4021  0.7602 -0.5523  0.6690 -0.6107 -0.5659 -0.0635
#> 
#> Columns 10 to 18  0.5427  0.3499  0.2805 -0.3018 -0.2263 -0.2887 -0.0895  0.7141 -0.6013
#>   0.3118 -0.2627  0.0903  0.3337  0.1271 -0.1767  0.1545 -0.0914 -0.2146
#>   0.0211 -0.3282  0.4454 -0.1623  0.3372  0.1571 -0.0892  0.1848  0.7414
#> 
#> Columns 19 to 20 -0.2560  0.0258
#>   0.1587 -0.3791
#>  -0.1045  0.0108
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6090  0.4113  0.1260  0.5464  0.2894  0.1881  0.3801 -0.0347 -0.1713
#>   0.3899  0.3530  0.1831  0.3906 -0.0715  0.2604  0.2271  0.2825 -0.2734
#>   0.0627  0.1209  0.4736  0.2941  0.2488  0.3566 -0.1929  0.5119 -0.3028
#> 
#> Columns 10 to 18 -0.1856 -0.3681 -0.4018  0.0685  0.0070 -0.1600  0.0591 -0.4705 -0.1571
#>  -0.3902 -0.3895 -0.1180 -0.2266  0.0014 -0.2404 -0.0752 -0.1714 -0.0779
#>  -0.5516 -0.3690 -0.4752 -0.1477  0.0560  0.0634  0.2268 -0.3784  0.0333
#> 
#> Columns 19 to 20  0.5690 -0.1477
#>   0.5435 -0.2648
#>   0.8181 -0.2728
#> [ CPUFloatType{2,3,20} ]
#>