RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6089  0.4602  0.2819 -0.1035  0.4753 -0.5779  0.9043 -0.3138 -0.3184
#>  -0.1668  0.9107 -0.3389 -0.7584  0.8043 -0.8653  0.9567  0.3708  0.2247
#>   0.6746 -0.5084  0.1129  0.0887  0.2846 -0.4604  0.6590 -0.2061 -0.5329
#> 
#> Columns 10 to 18 -0.8912  0.1762  0.0641 -0.4000  0.7348  0.0509  0.4547 -0.7444 -0.4266
#>  -0.0666  0.6485  0.0936 -0.5944  0.6715  0.8344 -0.3601  0.1305  0.2151
#>   0.4665 -0.1544  0.3037 -0.7348  0.8469  0.2937 -0.1322 -0.2109  0.5240
#> 
#> Columns 19 to 20 -0.2215  0.9120
#>   0.6791 -0.3558
#>   0.5989 -0.1553
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5451 -0.5056  0.2961  0.2246 -0.0977 -0.2302 -0.0032 -0.1765  0.4853
#>  -0.0946 -0.3072 -0.2956  0.3757 -0.0956 -0.4515  0.0456  0.3473  0.1285
#>   0.0536 -0.0520  0.0531  0.0074 -0.1282 -0.6287  0.1479 -0.4255  0.1399
#> 
#> Columns 10 to 18  0.4019 -0.6247  0.4785 -0.0097 -0.1761  0.1229 -0.0577 -0.3531  0.0219
#>   0.1482 -0.0424  0.4267 -0.2941  0.1338 -0.2272  0.7109  0.3736 -0.0082
#>  -0.1792 -0.1565  0.4888 -0.0333  0.1905  0.1829  0.0079 -0.1488  0.2861
#> 
#> Columns 19 to 20 -0.0816  0.3388
#>  -0.0242  0.5517
#>   0.5982  0.4324
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1134 -0.1657  0.0990  0.4410  0.0920 -0.5074  0.2939 -0.3682  0.0128
#>  -0.1807  0.1529  0.0075 -0.0791 -0.0994 -0.5090  0.2826 -0.5553  0.3054
#>  -0.3287  0.2993 -0.0284  0.0149 -0.0820 -0.0556  0.1231 -0.2570  0.2893
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1015 -0.3980  0.6561 -0.5477 -0.1555 -0.3988  0.6672  0.4899  0.2911
#>  -0.5610  0.5366 -0.5447  0.5876 -0.5855 -0.2837  0.5674  0.7112  0.0135
#>  -0.4399 -0.0984  0.8020 -0.1718 -0.1279  0.2467 -0.0640  0.4173  0.0044
#> 
#> Columns 10 to 18 -0.3667  0.3359  0.1736 -0.2001  0.4705  0.3678  0.5840 -0.1118  0.1112
#>  -0.1052 -0.8548  0.1006 -0.2218  0.0602 -0.6868 -0.2401  0.0698 -0.4450
#>  -0.5631 -0.0921  0.3781 -0.5473  0.9313  0.0646 -0.5100  0.2468  0.3032
#> 
#> Columns 19 to 20  0.1075 -0.6261
#>   0.0293  0.8377
#>  -0.5938  0.5563
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1062 -0.0159 -0.5936 -0.1365 -0.0592 -0.5104  0.3021 -0.1915  0.3874
#>  -0.2600  0.2933  0.2607 -0.1070 -0.4049 -0.2138  0.4095 -0.3198  0.4696
#>  -0.1636 -0.1092 -0.5170  0.0914 -0.0215 -0.1543  0.2151  0.3675  0.1864
#> 
#> Columns 10 to 18  0.1956  0.1923  0.7034  0.1348  0.1451  0.2856  0.2477 -0.0900  0.4370
#>  -0.0298  0.1274  0.5050 -0.0894  0.4294  0.4021 -0.1032  0.4484 -0.2047
#>  -0.0293  0.1158  0.6143 -0.1545  0.1137 -0.0332  0.0294  0.2798 -0.1031
#> 
#> Columns 19 to 20  0.4286  0.0037
#>  -0.2795  0.2768
#>   0.5254  0.1471
#> [ CPUFloatType{2,3,20} ]
#>