RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8241 -0.4075  0.2039 -0.7671  0.1954 -0.3497  0.2117  0.1853 -0.4011
#>   0.8958 -0.5375 -0.9329 -0.3223  0.4276 -0.7632  0.8826 -0.5107 -0.6939
#>  -0.4507  0.8429 -0.5968  0.5585 -0.4471  0.5613 -0.9508  0.8391  0.8661
#> 
#> Columns 10 to 18 -0.2230 -0.9276 -0.8579 -0.4404 -0.3756 -0.8513  0.4017 -0.7090 -0.9157
#>  -0.9851 -0.6538 -0.2447  0.7520 -0.4119 -0.7904  0.2972 -0.1346  0.0649
#>   0.3383 -0.7257 -0.9032  0.7007  0.0577 -0.8360  0.0797 -0.7521  0.0947
#> 
#> Columns 19 to 20  0.2309  0.6826
#>   0.1409  0.4411
#>   0.8260  0.2352
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.7225 -0.3313  0.1928  0.3714 -0.2213  0.4270  0.5923  0.3724 -0.1244
#>  -0.0394 -0.0024  0.4027  0.4733  0.3841 -0.0670  0.7538 -0.0538 -0.3322
#>   0.2756  0.6740  0.3707  0.2614 -0.0013 -0.2028  0.8404 -0.4571 -0.4638
#> 
#> Columns 10 to 18 -0.5614 -0.4040 -0.1765  0.3460  0.1833 -0.4077  0.1376 -0.1730 -0.1388
#>  -0.2585 -0.1731  0.3302  0.1004 -0.7156  0.1202  0.6846  0.1339 -0.5472
#>   0.2032  0.3136 -0.6924  0.6903  0.0905 -0.0632  0.0042 -0.3989  0.5432
#> 
#> Columns 19 to 20  0.4576 -0.0055
#>   0.6144 -0.3253
#>   0.6477 -0.1091
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3012  0.0550  0.4143  0.5196  0.3006  0.4404  0.4412 -0.0517 -0.4278
#>   0.1922 -0.0600  0.5296  0.2785  0.1908  0.3163  0.5005  0.2867 -0.4408
#>   0.2473  0.2806  0.3979  0.1534 -0.0627  0.5166 -0.2268  0.4075  0.3781
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2780  0.3003 -0.3187  0.2926 -0.3122 -0.1406  0.0671 -0.1259 -0.3097
#>  -0.1354  0.2719  0.1424 -0.2557  0.5906 -0.4878  0.3773  0.5415 -0.4224
#>  -0.4883  0.1408 -0.4905  0.3176  0.7663  0.3269  0.3323  0.3237 -0.0121
#> 
#> Columns 10 to 18  0.2069 -0.3381 -0.5193  0.3252 -0.4964 -0.2647 -0.2523  0.6690 -0.5555
#>  -0.2171  0.0481  0.1722  0.7754 -0.5649 -0.4073 -0.4908  0.0897 -0.0989
#>   0.0412 -0.2268  0.0529  0.1166 -0.4634 -0.4485 -0.3319 -0.2854 -0.5647
#> 
#> Columns 19 to 20 -0.3876 -0.3447
#>  -0.3058 -0.6303
#>  -0.1805  0.1533
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6707  0.2307  0.3761  0.3156 -0.3956  0.3777  0.5397  0.4623 -0.1511
#>   0.4891  0.1656 -0.0325  0.1914 -0.2100  0.3230  0.4900  0.1936 -0.1525
#>   0.5303  0.2335  0.2637  0.3607 -0.1638  0.5005  0.1322  0.4531  0.0666
#> 
#> Columns 10 to 18 -0.2305 -0.3029 -0.2468 -0.2307  0.1204 -0.1350 -0.0166 -0.1294 -0.3811
#>   0.0812  0.0957 -0.0751 -0.4447 -0.0874 -0.2229 -0.0007 -0.1836 -0.0750
#>  -0.2957 -0.3443 -0.3512  0.0332  0.4392 -0.2800 -0.0989 -0.4616 -0.0152
#> 
#> Columns 19 to 20  0.4795  0.2056
#>   0.6104  0.0622
#>   0.4640  0.0046
#> [ CPUFloatType{2,3,20} ]
#>