RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2068  0.8123  0.0582  0.1696  0.3793 -0.1548  0.1744  0.5112 -0.5230
#>  -0.0492 -0.2919  0.1651 -0.2145  0.5410  0.5104  0.1533 -0.6122 -0.4515
#>   0.1784 -0.4053  0.1501  0.4601  0.5169 -0.3951  0.2471  0.8004  0.5981
#> 
#> Columns 10 to 18 -0.6947  0.1261  0.2585  0.9406 -0.0285  0.1573 -0.1523  0.4327  0.6311
#>   0.2193 -0.1996  0.5506  0.9027  0.1660 -0.4124  0.3833  0.4876  0.5758
#>  -0.2803 -0.1193 -0.4161 -0.4863 -0.4420 -0.1336  0.2199 -0.5353  0.4226
#> 
#> Columns 19 to 20 -0.7652 -0.6275
#>   0.2168 -0.5955
#>  -0.8768 -0.0122
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2536  0.2599 -0.4656 -0.4385  0.2883  0.6266 -0.6253  0.0877  0.1414
#>  -0.1176  0.5051 -0.1791 -0.6035  0.3849  0.3604 -0.6390 -0.3299 -0.1875
#>  -0.3107 -0.3313 -0.5593  0.0837 -0.4575 -0.1240 -0.3526  0.0543  0.4240
#> 
#> Columns 10 to 18 -0.1335 -0.2367  0.1252  0.5489  0.6062 -0.2250 -0.5118  0.9126 -0.2652
#>  -0.0093 -0.0849  0.4391  0.2445  0.6711 -0.0196 -0.1553  0.7874  0.4452
#>  -0.0114  0.0983 -0.2245  0.3489  0.3060  0.0575  0.1011  0.1732  0.6107
#> 
#> Columns 19 to 20  0.2467 -0.3810
#>   0.0996 -0.3622
#>  -0.1763 -0.5174
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2475 -0.0680 -0.0639  0.0252  0.5173  0.2750 -0.3208  0.0543 -0.1863
#>  -0.5862  0.2306 -0.4403 -0.3294  0.0480  0.5077 -0.5416 -0.0064  0.3100
#>  -0.2518 -0.0015 -0.0017 -0.4956  0.0878  0.2662 -0.0531 -0.0183  0.3235
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1011  0.6102 -0.2724 -0.6707  0.4272  0.0315 -0.6190  0.0279  0.3750
#>   0.7257  0.9030 -0.0850  0.3662  0.4351  0.0529 -0.1491 -0.0242  0.6198
#>   0.2818 -0.4284  0.1609 -0.8241  0.4161  0.4475 -0.8402 -0.3128  0.7204
#> 
#> Columns 10 to 18  0.7162  0.5242 -0.2449  0.1819 -0.4517 -0.1086 -0.3421  0.4855 -0.7316
#>   0.8121  0.2111 -0.8194 -0.4475 -0.8476 -0.5720 -0.4454  0.4494 -0.7638
#>  -0.1401  0.1920  0.1120  0.0847 -0.5858 -0.7627  0.1853  0.3037 -0.7178
#> 
#> Columns 19 to 20 -0.6476  0.0214
#>  -0.2547 -0.7331
#>  -0.6806 -0.4754
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3616 -0.3408 -0.3555 -0.1993  0.0513  0.3997 -0.3283 -0.1855  0.3024
#>  -0.0185 -0.2566 -0.4933  0.1105 -0.2718  0.3811 -0.2879 -0.0105  0.0031
#>  -0.2705  0.0125 -0.4951 -0.3376 -0.2571  0.6319 -0.5314 -0.0627  0.1322
#> 
#> Columns 10 to 18 -0.0052  0.2853 -0.4378  0.4950  0.6626  0.0888 -0.0895  0.0699  0.4466
#>   0.4329  0.2469  0.0984  0.4257  0.8100 -0.3178 -0.6017 -0.2631  0.7576
#>   0.3266 -0.0429  0.4012  0.3712  0.4424 -0.3626 -0.0278  0.0936  0.5185
#> 
#> Columns 19 to 20  0.2744  0.0134
#>   0.0585  0.1831
#>   0.4218 -0.0954
#> [ CPUFloatType{2,3,20} ]
#>