RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1543  0.3728 -0.7216  0.0312  0.7347  0.0619  0.0768 -0.4057  0.4835
#>  -0.3781  0.4365  0.8397  0.4681 -0.0992 -0.8021 -0.5532  0.8502 -0.0393
#>  -0.2338 -0.5501  0.0994 -0.6966  0.6541 -0.7927  0.0982 -0.6166  0.1741
#> 
#> Columns 10 to 18 -0.6359 -0.0239  0.0501 -0.1236  0.1660  0.0917 -0.7402  0.3583 -0.6539
#>  -0.2839  0.4430  0.1765  0.5279 -0.7122 -0.6442 -0.3334  0.0593 -0.0321
#>   0.4715 -0.2942 -0.5387  0.3167  0.4344  0.5652 -0.2717  0.8535 -0.8043
#> 
#> Columns 19 to 20 -0.2496 -0.8787
#>   0.3824  0.3555
#>  -0.8995 -0.9720
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0399 -0.4769 -0.2091  0.3369  0.4437 -0.1532 -0.0133  0.2759 -0.3353
#>  -0.1420 -0.0139 -0.0733  0.2747  0.3778 -0.0833  0.0597 -0.0193  0.1901
#>   0.4509 -0.2455  0.0695  0.2322 -0.0843  0.4377 -0.2574  0.2644 -0.5721
#> 
#> Columns 10 to 18 -0.5008  0.0672 -0.2054 -0.0650  0.3125  0.2145 -0.2183 -0.1041 -0.4827
#>   0.0712 -0.1530 -0.2386 -0.1703 -0.3431 -0.4505  0.0301  0.2888 -0.3502
#>   0.0720 -0.3059 -0.1077 -0.1095  0.0882 -0.3953  0.1905 -0.1265 -0.7940
#> 
#> Columns 19 to 20 -0.1634 -0.0846
#>  -0.4980  0.3737
#>  -0.5386  0.6407
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2424  0.0089  0.1912 -0.1794  0.2154  0.0456 -0.2527  0.3395 -0.3305
#>  -0.2998  0.1915 -0.2645 -0.0709  0.3349  0.2218 -0.2513  0.2196 -0.1525
#>  -0.3334  0.1591  0.0974  0.1205 -0.0974  0.2959 -0.5265  0.6177 -0.0173
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8360  0.7873 -0.3303 -0.2181 -0.5529 -0.1731  0.2073  0.4828 -0.0850
#>   0.0196  0.1058 -0.3942 -0.0358  0.3845  0.5394 -0.2918  0.4097  0.2430
#>   0.5094 -0.4422 -0.6363  0.6502  0.3462 -0.6095 -0.1938 -0.8228  0.2807
#> 
#> Columns 10 to 18  0.2779  0.4582 -0.3870 -0.1238  0.2122  0.2501 -0.5680  0.0440  0.5817
#>   0.3642  0.5519 -0.7422 -0.0084 -0.7134  0.7165  0.0446  0.2144  0.1778
#>   0.3608  0.5024 -0.8229 -0.0426 -0.1242  0.8098  0.0327  0.6452 -0.6719
#> 
#> Columns 19 to 20  0.1353 -0.1794
#>  -0.4352  0.2053
#>   0.3410  0.3758
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0345  0.0827  0.1191 -0.0068  0.0096  0.3597 -0.4453  0.2626 -0.3751
#>   0.0388  0.3755  0.0610  0.1559  0.2661  0.2691 -0.2199  0.1242 -0.2792
#>   0.0292  0.0627 -0.3460  0.3233  0.3755 -0.0223 -0.0798  0.1922 -0.2791
#> 
#> Columns 10 to 18 -0.3314 -0.0120 -0.1695 -0.4333 -0.1842 -0.1189  0.0416 -0.1829 -0.2624
#>  -0.3948 -0.1079  0.2421  0.0137 -0.4406 -0.1379  0.1237  0.4619 -0.3286
#>  -0.8112 -0.0015  0.1492  0.0390  0.0974 -0.1227  0.1409  0.3664 -0.4405
#> 
#> Columns 19 to 20 -0.3774  0.5296
#>  -0.3888  0.3316
#>  -0.4039  0.2417
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>