RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7000  0.7326  0.8373  0.0372  0.8928 -0.6812  0.3742 -0.2042 -0.0840
#>   0.1661  0.9018 -0.2580 -0.1412  0.8441  0.2125  0.4635 -0.5868 -0.4531
#>   0.0353 -0.5240  0.3644 -0.7722  0.7689  0.5025  0.5735  0.7586 -0.8575
#> 
#> Columns 10 to 18  0.0152 -0.6106 -0.0320 -0.0238  0.4720  0.6600  0.5196 -0.4979  0.2964
#>  -0.4977  0.6793 -0.2281  0.6740  0.6265 -0.3696  0.1045 -0.6200 -0.9067
#>  -0.0414  0.3727 -0.3009  0.8039  0.1124  0.0134  0.5186 -0.1112 -0.8299
#> 
#> Columns 19 to 20 -0.2622  0.2178
#>   0.6308 -0.5425
#>  -0.5443 -0.3228
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3800  0.4521 -0.1070  0.4019  0.3720  0.4332  0.0898 -0.3833 -0.1103
#>  -0.2656  0.4067  0.5303 -0.0436 -0.0615 -0.4755 -0.1555 -0.4777  0.1228
#>  -0.3690  0.7228 -0.2864  0.2563  0.3258 -0.3838 -0.2995 -0.0282 -0.5653
#> 
#> Columns 10 to 18  0.0284  0.0821  0.0432 -0.6221  0.2416  0.0837 -0.1375 -0.1773 -0.0571
#>   0.1374  0.1875  0.2780 -0.1896 -0.3872  0.5353  0.6281 -0.6535 -0.5315
#>  -0.2489  0.0373  0.0629  0.2236  0.5577  0.6162  0.4850 -0.4703 -0.6936
#> 
#> Columns 19 to 20  0.2872 -0.5123
#>   0.4750  0.1302
#>  -0.4072  0.3313
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2286  0.2600  0.4619  0.3393  0.1925  0.0348  0.0118 -0.1363  0.2767
#>  -0.6464  0.1791 -0.2037  0.1151 -0.4044  0.1493  0.4963 -0.5379  0.3364
#>  -0.1067  0.4419 -0.2006  0.3727  0.1173  0.4056 -0.0722  0.0719 -0.3169
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2614  0.2039 -0.1484  0.2627 -0.6281  0.7927 -0.0038 -0.1539  0.3562
#>  -0.6124 -0.4416 -0.6038  0.7167  0.5729 -0.5218  0.6230  0.2944 -0.5338
#>   0.2744 -0.4477  0.2714 -0.2036 -0.1037  0.1199  0.1678 -0.5142  0.4766
#> 
#> Columns 10 to 18  0.0339  0.3770 -0.2607  0.6226  0.2296 -0.7191 -0.7283 -0.3405  0.2657
#>  -0.1024 -0.6455  0.2939  0.2101  0.1693  0.0227  0.0851 -0.8357 -0.2388
#>  -0.4755  0.1550  0.2274 -0.0710  0.5862  0.0154  0.2258 -0.0079 -0.3506
#> 
#> Columns 19 to 20  0.5875 -0.0992
#>   0.1789  0.1126
#>   0.2760  0.0041
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3906  0.2600  0.0172 -0.1995  0.0271  0.5623  0.0464  0.2309  0.1023
#>  -0.2822  0.0452 -0.0127 -0.0345 -0.3958  0.1557  0.7406 -0.6179  0.3989
#>  -0.2115  0.3913  0.1271 -0.3547 -0.1540 -0.2147  0.0567  0.3949 -0.0954
#> 
#> Columns 10 to 18 -0.1915 -0.1296 -0.2708  0.1166  0.4109  0.0015 -0.2402 -0.1888 -0.2555
#>  -0.0901 -0.0407  0.2385 -0.0030  0.0761 -0.0739  0.1271 -0.5330  0.2588
#>   0.1178 -0.0821  0.3681 -0.2821  0.1036  0.1611  0.0928 -0.2539 -0.5149
#> 
#> Columns 19 to 20  0.0687  0.3765
#>   0.0025  0.2354
#>   0.2550  0.4527
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>