RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8421 -0.3440  0.1765 -0.4741 -0.3917 -0.0677  0.1847  0.7005  0.4963
#>   0.0200 -0.1831 -0.8985  0.0252  0.1470  0.4791 -0.1484 -0.3156  0.5668
#>  -0.0565 -0.8254  0.5548 -0.2114  0.3436  0.4766 -0.4153 -0.2283  0.4255
#> 
#> Columns 10 to 18  0.6163 -0.2294 -0.7820  0.4285 -0.2781  0.2649  0.6158  0.0733 -0.0799
#>  -0.4248 -0.3689  0.1772  0.7826  0.8010  0.3438  0.3226 -0.7377 -0.3370
#>  -0.1083 -0.3492 -0.6229  0.8312  0.0738  0.3493  0.2440 -0.7211  0.6678
#> 
#> Columns 19 to 20  0.7664  0.0787
#>  -0.5028 -0.4023
#>   0.2500 -0.3570
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1275 -0.4367 -0.5634 -0.5838  0.2834  0.3771  0.3022 -0.4053 -0.2539
#>   0.0578 -0.6716 -0.4346 -0.2184 -0.0002 -0.1148 -0.1647 -0.3190  0.3750
#>  -0.2464 -0.3508 -0.6325  0.1575  0.5951 -0.2622  0.0017 -0.3445 -0.0433
#> 
#> Columns 10 to 18 -0.1741  0.1512  0.2443  0.2796 -0.1248  0.0943  0.3233 -0.0390 -0.1916
#>  -0.2066  0.5325  0.5905  0.5810  0.5577  0.1890  0.7843  0.2975  0.4611
#>  -0.2407 -0.2831  0.4807  0.3636 -0.2627 -0.3768  0.8194  0.4794 -0.3081
#> 
#> Columns 19 to 20 -0.0495 -0.5039
#>   0.3222 -0.2050
#>   0.2305 -0.5001
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2571  0.1657 -0.5357  0.0232  0.1938 -0.1486 -0.0909 -0.3796  0.0708
#>  -0.5099 -0.4859  0.3973  0.1368  0.2201 -0.3106  0.2330 -0.2737  0.8864
#>   0.1118 -0.2423  0.3145  0.1948  0.2244  0.0695 -0.0737 -0.6428  0.2403
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3468 -0.2270 -0.4301  0.1145 -0.6697 -0.2091 -0.3309 -0.5602  0.0220
#>   0.1228 -0.1921 -0.1463 -0.4601  0.2513 -0.3440 -0.6835 -0.2868 -0.0780
#>  -0.3397  0.2624 -0.3011  0.5278 -0.0686 -0.2970 -0.6285 -0.7868 -0.4759
#> 
#> Columns 10 to 18  0.6041 -0.5746  0.4164 -0.2762 -0.6244  0.0229  0.5060  0.4614 -0.1953
#>   0.4870 -0.2453 -0.4138 -0.2672 -0.1526 -0.0529  0.4221  0.2255  0.0320
#>   0.2979 -0.1618 -0.0687 -0.3003 -0.5635 -0.7229  0.2371  0.2827  0.2174
#> 
#> Columns 19 to 20 -0.6827 -0.1550
#>  -0.1206  0.6386
#>  -0.6639 -0.5995
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5507 -0.5762  0.0900 -0.0379  0.3794 -0.4873 -0.0316 -0.2691  0.5677
#>  -0.3595 -0.2276  0.3101  0.0349  0.3899 -0.3416  0.1778 -0.0193  0.6795
#>  -0.3015 -0.6662 -0.0642  0.3045  0.3761 -0.5247 -0.4052 -0.3730  0.1769
#> 
#> Columns 10 to 18  0.1346  0.4869  0.4578  0.3254  0.1714  0.1343  0.8396  0.1553 -0.1980
#>   0.0462  0.2428  0.1511  0.5589 -0.0192  0.2667  0.7008  0.4968 -0.1133
#>   0.1087  0.0072  0.4096  0.4294 -0.0246  0.1950  0.5818  0.6986 -0.1519
#> 
#> Columns 19 to 20 -0.2036  0.0541
#>   0.0137 -0.4150
#>  -0.1515 -0.4983
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>