RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5356  0.9390  0.4575  0.2926 -0.6266 -0.8437 -0.3807 -0.6979  0.0244
#>   0.8732 -0.3420 -0.0158 -0.5991 -0.3189 -0.4147  0.8362  0.8379 -0.4435
#>   0.8073 -0.3483  0.2006  0.5832 -0.9256 -0.5348  0.5166 -0.2668  0.7052
#> 
#> Columns 10 to 18  0.4526  0.5478  0.6033  0.0681  0.8035  0.4807 -0.4172 -0.5835  0.8056
#>   0.0155 -0.7083 -0.8210  0.0214 -0.7776 -0.2719  0.7802 -0.0338 -0.1976
#>   0.6267  0.5345 -0.4852 -0.2108 -0.3112 -0.1335 -0.2219  0.4101  0.4744
#> 
#> Columns 19 to 20 -0.4318 -0.8115
#>  -0.0850  0.6634
#>   0.2492 -0.0781
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4979  0.5677  0.5223 -0.4605 -0.2093 -0.6733 -0.1949 -0.5264 -0.3993
#>   0.4257 -0.5378  0.1121  0.0136 -0.7344 -0.1076 -0.2754  0.6282  0.0635
#>  -0.1618  0.2276  0.1882  0.2025 -0.1945 -0.8207 -0.2863 -0.5980 -0.0938
#> 
#> Columns 10 to 18  0.5272 -0.2520  0.1470 -0.5104 -0.3064 -0.6418  0.7428 -0.6722 -0.3858
#>   0.0665  0.2357 -0.8059 -0.4285 -0.3635 -0.6381  0.6812  0.4467  0.3513
#>   0.1564  0.0014  0.4940  0.0116 -0.1721 -0.1069  0.1526 -0.3430  0.6108
#> 
#> Columns 19 to 20  0.0299 -0.1625
#>   0.4472  0.0982
#>  -0.5616  0.0598
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.7007 -0.0954  0.2358 -0.3124 -0.4806 -0.4358 -0.1234  0.0251 -0.1165
#>   0.1414  0.4221  0.3677  0.3483 -0.5073 -0.4837 -0.1990  0.2212  0.2502
#>   0.6632 -0.3817  0.3330 -0.5206 -0.2567 -0.2120  0.2679 -0.0960 -0.0666
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2768 -0.4054 -0.5011  0.4103 -0.6814 -0.3250 -0.6562 -0.2996  0.3215
#>   0.4453 -0.2877  0.6769 -0.2091 -0.5944  0.4609 -0.4768  0.3023  0.1801
#>   0.5047 -0.5278  0.0932  0.4275 -0.1799  0.5975  0.1915  0.1403  0.2835
#> 
#> Columns 10 to 18 -0.2959  0.1039 -0.1838  0.1996 -0.4412 -0.0264  0.2080 -0.0026 -0.5765
#>  -0.2451  0.3706 -0.1117 -0.9301  0.8649  0.3708 -0.0768  0.4399  0.2265
#>  -0.0810  0.2628 -0.3694 -0.5540  0.9544 -0.6209  0.3139  0.7206 -0.3660
#> 
#> Columns 19 to 20 -0.2957 -0.2171
#>  -0.1852 -0.5684
#>  -0.0978 -0.1508
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2425 -0.0547  0.2157 -0.4134 -0.3301 -0.3003  0.1800  0.2068  0.2779
#>   0.2586  0.4143 -0.4274 -0.4003 -0.2295 -0.1772  0.2376 -0.2009  0.0636
#>   0.7268  0.0067  0.2340 -0.2493 -0.4798 -0.2673  0.1423 -0.4716 -0.4529
#> 
#> Columns 10 to 18 -0.0524  0.2812 -0.1537 -0.0461 -0.1436  0.0418  0.0453 -0.0366  0.3087
#>   0.3329 -0.1227 -0.2097 -0.1557  0.1289 -0.3957  0.3015  0.1066 -0.2258
#>   0.4825  0.0784 -0.4631 -0.2075 -0.5058 -0.6757  0.5046 -0.5571 -0.2793
#> 
#> Columns 19 to 20  0.2397 -0.0329
#>  -0.3988  0.3398
#>   0.0380  0.0932
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>