RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4384  0.7721  0.6569  0.5279  0.7406 -0.0490 -0.1584 -0.5404 -0.7677
#>   0.6955 -0.3591  0.6266 -0.1807  0.1769  0.0797  0.5767 -0.0517 -0.3665
#>   0.0175 -0.0267  0.7502  0.9209  0.2381 -0.0094  0.2192  0.1475 -0.2767
#> 
#> Columns 10 to 18  0.4447  0.1903 -0.8262  0.0483 -0.5055  0.6116 -0.4059 -0.1009  0.7865
#>  -0.8233  0.3064  0.5186  0.2385  0.4385  0.8471  0.6060 -0.0378 -0.5913
#>   0.5902 -0.0662 -0.3432 -0.3479 -0.6230  0.2457 -0.1130  0.1031  0.0805
#> 
#> Columns 19 to 20  0.7409 -0.5677
#>  -0.2057  0.2954
#>   0.9583 -0.8298
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2922  0.0870  0.4172 -0.2459 -0.1282  0.0021  0.3348  0.1951 -0.2225
#>   0.4713 -0.0768  0.0829  0.3930  0.0405  0.5731 -0.0821  0.7314  0.2762
#>   0.5584  0.2168  0.4403  0.0545 -0.0133 -0.3059  0.7923 -0.3078 -0.0954
#> 
#> Columns 10 to 18 -0.0870 -0.2746 -0.2345 -0.5843  0.5009 -0.2738  0.0737 -0.1716 -0.3700
#>   0.0372 -0.1489  0.0160 -0.1469  0.0274  0.0653  0.5164 -0.4450 -0.7776
#>  -0.4002 -0.2766  0.0490 -0.5788  0.2386 -0.0385 -0.1746  0.2306 -0.0613
#> 
#> Columns 19 to 20  0.1956  0.0644
#>   0.4224  0.2128
#>   0.3899  0.2987
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5112  0.2080  0.2213  0.4355 -0.3526  0.4227  0.7016  0.5468 -0.3839
#>   0.3252  0.3920  0.4195  0.1844  0.0259  0.2220  0.0437  0.4674 -0.0746
#>   0.1882  0.3055  0.2801  0.2276 -0.2856  0.5586  0.0025  0.8567 -0.0912
#> 
#> Columns 10 to 18 -0.1230 -0.1833 -0.2684  0.3406  0.1829  0.0789  0.3191 -0.2895 -0.2948
#>  -0.7310 -0.2210 -0.4205  0.2787  0.3304 -0.1164  0.1440 -0.3412  0.1418
#>  -0.2971 -0.0803 -0.0206 -0.4882  0.0101 -0.0134  0.3407 -0.0893 -0.2795
#> 
#> Columns 19 to 20  0.5031  0.1467
#>   0.6732 -0.0413
#>   0.7675 -0.0210
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.5865  0.4415 -0.1009  0.3233 -0.1824  0.5614  0.0790  0.6864  0.0418
#>   0.3474 -0.1574  0.5494  0.7235  0.1103  0.3606  0.3251  0.1986 -0.0627
#>   0.6635  0.3085  0.3537  0.3851  0.1092  0.1799  0.4390  0.1624 -0.4746
#> 
#> Columns 10 to 18 -0.4047 -0.4346 -0.2003  0.0779  0.2632 -0.2331  0.1536 -0.4161 -0.3587
#>  -0.1713 -0.3531 -0.3912  0.2909  0.4168  0.1489 -0.0635 -0.3876 -0.0038
#>  -0.2874 -0.2212 -0.5173  0.2847  0.0833  0.2829  0.1235 -0.6485 -0.0689
#> 
#> Columns 19 to 20  0.4875  0.1354
#>   0.3792 -0.0990
#>   0.5475 -0.0600
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.2711  0.2275  0.3727  0.3354 -0.0777  0.4432  0.4550  0.2527 -0.3596
#>   0.1524  0.1214  0.3465  0.4097 -0.0545  0.4506  0.5084 -0.0077 -0.2751
#>   0.2882  0.1242  0.4013  0.1809 -0.2909  0.5230  0.0735  0.5625  0.1028
#> 
#> Columns 10 to 18 -0.3900 -0.3703 -0.4742  0.4051  0.1727 -0.0744  0.2091 -0.4474 -0.1320
#>  -0.4472 -0.5057 -0.3773 -0.1574  0.2564 -0.3629 -0.0371  0.0844 -0.0814
#>  -0.4835 -0.4336 -0.1385  0.0089  0.3320 -0.2400  0.0718 -0.2691 -0.2223
#> 
#> Columns 19 to 20  0.6364 -0.3889
#>   0.4919 -0.2789
#>   0.4929  0.1467
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0417 -0.3994 -0.5907  0.8831  0.1489  0.3421 -0.3280 -0.5027 -0.3500
#>  -0.4876 -0.0346 -0.3219  0.5960 -0.5579  0.6480 -0.5114 -0.1279 -0.4376
#>  -0.3691  0.4150 -0.5231  0.3444  0.3943  0.3629 -0.1251 -0.1202 -0.5187
#> 
#> Columns 10 to 18  0.3859 -0.0414 -0.1612  0.4102 -0.3215 -0.1030 -0.6172  0.2624  0.0453
#>   0.8229  0.1933 -0.2790 -0.1263 -0.2762 -0.5762 -0.2625  0.3174  0.1282
#>  -0.2068 -0.2465  0.0576 -0.0155 -0.3770 -0.4816 -0.4269  0.0079 -0.4219
#> 
#> Columns 19 to 20 -0.4129  0.2686
#>  -0.4658  0.2666
#>   0.1670 -0.2883
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2711  0.2275  0.3727  0.3354 -0.0777  0.4432  0.4550  0.2527 -0.3596
#>   0.1524  0.1214  0.3465  0.4097 -0.0545  0.4506  0.5084 -0.0077 -0.2751
#>   0.2882  0.1242  0.4013  0.1809 -0.2909  0.5230  0.0735  0.5625  0.1028
#> 
#> Columns 10 to 18 -0.3900 -0.3703 -0.4742  0.4051  0.1727 -0.0744  0.2091 -0.4474 -0.1320
#>  -0.4472 -0.5057 -0.3773 -0.1574  0.2564 -0.3629 -0.0371  0.0844 -0.0814
#>  -0.4835 -0.4336 -0.1385  0.0089  0.3320 -0.2400  0.0718 -0.2691 -0.2223
#> 
#> Columns 19 to 20  0.6364 -0.3889
#>   0.4919 -0.2789
#>   0.4929  0.1467
#> [ CPUFloatType{2,3,20} ]
#>