RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1479 -0.6440  0.2337  0.2944  0.0179  0.4720  0.4986 -0.7661  0.1629
#>   0.1709 -0.5384 -0.2263 -0.2017 -0.2873  0.3416  0.3542 -0.4816  0.3782
#>   0.4808 -0.2532  0.3522 -0.6107  0.3129 -0.1423 -0.8487 -0.4255 -0.5052
#> 
#> Columns 10 to 18 -0.6320 -0.4925  0.5546 -0.3902 -0.6275 -0.0070 -0.1749 -0.2532 -0.3664
#>   0.4233 -0.0505  0.2173  0.5011 -0.4747  0.1978 -0.1531  0.0770  0.8327
#>   0.7773  0.6903  0.2974  0.7872 -0.6825  0.5132  0.2897  0.6047 -0.5950
#> 
#> Columns 19 to 20 -0.0369 -0.6309
#>  -0.0615 -0.7090
#>   0.3995 -0.0861
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1014 -0.1229 -0.2335  0.3131 -0.2330 -0.0066  0.1926  0.3366  0.3338
#>  -0.2134  0.1267  0.2451  0.0624 -0.3740 -0.0063 -0.4011  0.1447 -0.1073
#>  -0.0538 -0.2976  0.4104 -0.6320 -0.3519  0.1169 -0.3989  0.3504  0.4205
#> 
#> Columns 10 to 18 -0.2890 -0.3323  0.3417  0.1512 -0.3485  0.2571  0.0633  0.3109 -0.1866
#>  -0.5057  0.2612  0.4143 -0.2830 -0.3297  0.4315 -0.2985  0.2521  0.3563
#>   0.1288 -0.3048  0.2834  0.0428  0.1213  0.5206 -0.0157  0.5102  0.4054
#> 
#> Columns 19 to 20 -0.3083  0.1908
#>  -0.1018 -0.1965
#>  -0.7548 -0.3727
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2818  0.0706  0.0752  0.6619 -0.1926  0.2363 -0.2634 -0.1662 -0.1959
#>  -0.1206  0.2454 -0.1454  0.2987 -0.2471 -0.0202  0.0566  0.1902 -0.4848
#>  -0.3582  0.1272 -0.4712  0.3007 -0.3966  0.0400 -0.2635  0.0217 -0.5845
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1239 -0.1814  0.0475 -0.2100  0.1886  0.7784 -0.5331 -0.4532  0.5177
#>   0.3549  0.4696 -0.0309  0.7638 -0.0140  0.4388 -0.6917 -0.5989  0.3694
#>  -0.3442  0.3769  0.4032  0.1902 -0.0800  0.0197 -0.4105  0.0505  0.2845
#> 
#> Columns 10 to 18  0.1514  0.4355  0.4928 -0.8228 -0.3185  0.4788 -0.3632 -0.6055 -0.1541
#>   0.3043 -0.7799  0.1768 -0.2213 -0.3588  0.0272  0.8229 -0.4099  0.2689
#>   0.2885  0.1199  0.2820 -0.4688 -0.3546 -0.3539  0.0264 -0.5000  0.2615
#> 
#> Columns 19 to 20 -0.6241  0.4331
#>   0.4749  0.2265
#>  -0.5256 -0.2964
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0067  0.0942 -0.0077  0.3233 -0.2450  0.3545  0.1763 -0.2319  0.0134
#>  -0.2998  0.0330  0.1040  0.5211 -0.5765  0.2847 -0.4136 -0.4076 -0.2876
#>  -0.3052  0.1396  0.0173  0.2041 -0.3066  0.1194 -0.1290 -0.0051 -0.1018
#> 
#> Columns 10 to 18 -0.2280 -0.2254  0.4627  0.1592 -0.3574  0.4737  0.1665  0.4650  0.1481
#>  -0.1212  0.4531  0.8005  0.0405 -0.3397  0.4999 -0.2777  0.5675  0.2019
#>  -0.2673 -0.3802  0.2844  0.3904 -0.4912  0.6767  0.0390  0.2841  0.2524
#> 
#> Columns 19 to 20 -0.3664 -0.2022
#>   0.0546 -0.4724
#>  -0.5821 -0.1916
#> [ CPUFloatType{2,3,20} ]
#>