RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0709  0.2395  0.3089 -0.0877  0.8642  0.2254 -0.4456  0.6288 -0.3583
#>  -0.4019 -0.1378 -0.6290 -0.5746  0.3585 -0.5598 -0.0962 -0.0295  0.1281
#>   0.1496 -0.4617 -0.1243 -0.4531  0.4018 -0.3185 -0.3552 -0.4454  0.4882
#> 
#> Columns 10 to 18 -0.7023  0.6255 -0.2177  0.1208  0.4086  0.9025  0.1224  0.2494  0.1307
#>  -0.4758  0.5827 -0.2056  0.5556 -0.2881  0.3380  0.4130  0.9656 -0.2981
#>  -0.4615  0.2002 -0.2700  0.1769  0.2162  0.2659  0.5309  0.8733 -0.0474
#> 
#> Columns 19 to 20  0.6783  0.7202
#>   0.4186 -0.4900
#>   0.4503  0.7986
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1789  0.3488 -0.3724 -0.6453  0.0180  0.7045 -0.3393  0.0347 -0.3228
#>   0.4008  0.2039 -0.7837 -0.2753  0.3790  0.7640  0.3323  0.7335 -0.5390
#>  -0.1517  0.5924 -0.4294  0.1054  0.5505  0.2702 -0.2286  0.3273 -0.5259
#> 
#> Columns 10 to 18  0.3109 -0.2416 -0.1120  0.1556 -0.6488  0.3403 -0.2449 -0.2526 -0.0854
#>   0.4598  0.5852 -0.4823 -0.3167 -0.2289 -0.0139  0.5117  0.1092  0.1415
#>   0.2284 -0.0932  0.3646  0.0120 -0.1543  0.2395  0.0473  0.1793  0.4464
#> 
#> Columns 19 to 20  0.0149  0.0439
#>  -0.3206  0.6841
#>  -0.1607 -0.0567
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0151  0.4250  0.0945 -0.3587  0.2554  0.2113 -0.1819  0.2168 -0.2747
#>   0.2564  0.6568 -0.1234 -0.3043  0.0989 -0.5539 -0.4291 -0.1705 -0.4505
#>   0.4437  0.2991  0.2875 -0.3601  0.5095  0.2600 -0.0011  0.1393  0.0815
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3806  0.5519 -0.2144  0.1060  0.0245 -0.1731 -0.1452  0.1725 -0.4361
#>   0.1974 -0.6197 -0.7426 -0.5358 -0.0663 -0.3258 -0.2499 -0.0289 -0.5484
#>  -0.6774 -0.6937 -0.2722 -0.6679  0.1780  0.7171  0.4292  0.5634 -0.5475
#> 
#> Columns 10 to 18  0.5740 -0.0847 -0.0523  0.1123  0.3826 -0.4937 -0.0755  0.6467  0.6273
#>   0.8224  0.6245  0.0130  0.8503 -0.8746 -0.5259  0.0530  0.7705  0.8702
#>   0.2009  0.1415 -0.3347  0.0722  0.6682 -0.3182  0.0198 -0.1786 -0.5414
#> 
#> Columns 19 to 20  0.1535  0.5133
#>  -0.2765  0.1189
#>  -0.6385  0.2981
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3977  0.3985 -0.0782 -0.0538  0.4489  0.3279  0.1230  0.3286 -0.1379
#>   0.3566  0.2444 -0.0415 -0.6195  0.5320  0.5558  0.0963  0.5650 -0.2650
#>  -0.0907  0.3116 -0.1173 -0.5049  0.4407 -0.1949 -0.3841  0.1937 -0.4470
#> 
#> Columns 10 to 18  0.1089 -0.0690 -0.2735 -0.1418  0.2342  0.4853  0.2907  0.3115  0.0593
#>   0.8710 -0.1575 -0.2239  0.0831 -0.3363 -0.0373  0.4660 -0.0654  0.1712
#>   0.0898  0.0971  0.2399 -0.1637 -0.6883  0.3530 -0.2861  0.1725 -0.3385
#> 
#> Columns 19 to 20  0.2577  0.1426
#>  -0.2538 -0.3919
#>   0.4058 -0.0156
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>