RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6879 -0.5156 -0.5937  0.5299  0.5878 -0.6984 -0.3394 -0.2197  0.0444
#>  -0.7921 -0.1600 -0.8611  0.9012  0.3289  0.2083 -0.3954  0.5359 -0.4102
#>   0.4416 -0.1897 -0.9202 -0.1364 -0.3886  0.4878 -0.4543  0.5196  0.7424
#> 
#> Columns 10 to 18 -0.0198 -0.2481 -0.4294 -0.4432  0.5067 -0.0669  0.5565  0.5206  0.3252
#>  -0.3116 -0.8029 -0.5426 -0.1020  0.8225 -0.5877 -0.6547 -0.0516  0.2642
#>  -0.6596  0.8319 -0.5986  0.4595 -0.3351  0.4914  0.1888  0.2298 -0.4412
#> 
#> Columns 19 to 20  0.6807 -0.4915
#>   0.0453  0.7212
#>  -0.3608  0.3065
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3076 -0.3485 -0.6541  0.2468 -0.6515  0.5311  0.1345 -0.0792  0.0872
#>   0.1295  0.1499 -0.3368 -0.1242 -0.3868  0.7269  0.7167 -0.6882  0.8425
#>  -0.0048 -0.4317 -0.0726  0.1769 -0.4441 -0.1362 -0.0920 -0.5011 -0.0587
#> 
#> Columns 10 to 18  0.3402 -0.1629  0.3486  0.2143 -0.0831 -0.4223  0.3912  0.3270 -0.0100
#>   0.2937  0.5362  0.5833  0.3118 -0.4719  0.3718  0.1778  0.1726  0.4104
#>  -0.6410  0.4349  0.0119  0.2186 -0.2900 -0.3975 -0.1371  0.3313 -0.0319
#> 
#> Columns 19 to 20  0.1568  0.5056
#>  -0.4515 -0.0413
#>   0.0553  0.2087
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1530 -0.2342 -0.4467  0.6439 -0.3029  0.3019 -0.3868  0.3462  0.0293
#>  -0.3633 -0.5182 -0.4273  0.6906  0.1027 -0.2139 -0.5352  0.2710 -0.5749
#>   0.4476 -0.6293 -0.2376  0.0268 -0.2870 -0.3494 -0.3162  0.1245 -0.1212
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5251  0.5693  0.5148 -0.4559 -0.8179  0.0035  0.7053 -0.1890 -0.4846
#>  -0.8199 -0.2788  0.6368  0.5763 -0.7426 -0.5750  0.2010 -0.0085 -0.2867
#>   0.2063  0.8241  0.3638  0.0117 -0.4909 -0.2095  0.6344 -0.4598  0.4633
#> 
#> Columns 10 to 18  0.3842 -0.7930 -0.0203 -0.3236  0.4124  0.7260  0.2445 -0.2928  0.0302
#>   0.5585 -0.6409 -0.3616  0.3076  0.6655 -0.2165  0.7492  0.0013 -0.3078
#>   0.5075 -0.6160  0.5318  0.1687  0.0363  0.3136 -0.4375  0.3638 -0.2693
#> 
#> Columns 19 to 20 -0.5801 -0.7650
#>   0.1703 -0.7545
#>  -0.6488  0.3466
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1680  0.0092 -0.0012 -0.1361 -0.5457 -0.2949  0.0966 -0.2151  0.3395
#>  -0.0119 -0.3906 -0.5938 -0.0715 -0.6676 -0.3264 -0.0586 -0.6010  0.4121
#>  -0.0798 -0.5360 -0.4642  0.2317 -0.3978 -0.1779 -0.4780 -0.1878 -0.0961
#> 
#> Columns 10 to 18 -0.3898  0.4649  0.3522  0.3493 -0.1200 -0.3298 -0.0328  0.0207  0.1496
#>  -0.3651  0.0974  0.5882  0.5379 -0.3078  0.4433  0.4153  0.4848  0.4672
#>  -0.6161 -0.0761  0.3026  0.2848 -0.2818  0.0311  0.3132 -0.1685 -0.3501
#> 
#> Columns 19 to 20  0.0428  0.7087
#>   0.2702  0.4530
#>  -0.3756  0.4321
#> [ CPUFloatType{2,3,20} ]
#>