RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8481  0.7789 -0.0292 -0.6504 -0.2007  0.4412  0.5083  0.1366 -0.9414
#>   0.0014  0.7268 -0.6239  0.2137 -0.8971 -0.3760  0.3446 -0.3724  0.0896
#>   0.2321  0.5211  0.7813 -0.7827  0.1726  0.3105 -0.5662  0.4461  0.4842
#> 
#> Columns 10 to 18  0.5960  0.7132  0.2218  0.6754  0.1890  0.1192 -0.3438 -0.5683  0.0288
#>   0.6875  0.7628 -0.3655 -0.4945 -0.7419 -0.9113  0.7366  0.9195  0.2169
#>   0.8545  0.3091 -0.0624 -0.8958 -0.0887  0.4956  0.6646 -0.8193 -0.4303
#> 
#> Columns 19 to 20  0.8817 -0.0055
#>   0.7544  0.3749
#>   0.7754 -0.1684
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5119 -0.1835 -0.4830 -0.2691  0.3399 -0.2852  0.7092  0.2553  0.7592
#>   0.1430  0.6111  0.2879  0.6208 -0.0167  0.3521 -0.1307 -0.5288 -0.0803
#>   0.2348  0.6204 -0.3692  0.4811  0.3434 -0.2380  0.6117 -0.1410 -0.3836
#> 
#> Columns 10 to 18  0.3865 -0.2338  0.5139  0.0207  0.0984  0.2884  0.2049  0.2848 -0.2686
#>   0.7079  0.1507  0.3693  0.0345 -0.1548 -0.5553 -0.2339  0.1805 -0.2620
#>   0.6193 -0.2174  0.0226  0.3131  0.5215 -0.5112  0.0919 -0.3629  0.0012
#> 
#> Columns 19 to 20  0.8165  0.5759
#>  -0.3600  0.6225
#>   0.8790 -0.1746
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1918  0.0029 -0.3612  0.0927 -0.0514  0.2224  0.2165  0.0691  0.4338
#>  -0.4576  0.7094 -0.0606  0.2781  0.2459  0.2493 -0.1313  0.0885 -0.2029
#>  -0.0454 -0.0746 -0.1735  0.1066  0.2099  0.6105 -0.1529 -0.1501  0.0734
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1604  0.0372  0.1188 -0.2274 -0.2835 -0.1917  0.7052  0.1269  0.2792
#>   0.0559  0.1859  0.2141  0.2804 -0.4316  0.4347 -0.3862 -0.5075  0.4429
#>  -0.1267  0.4180 -0.3468 -0.2120 -0.6564 -0.2843  0.0302 -0.0848  0.3659
#> 
#> Columns 10 to 18  0.3048  0.1275  0.4040 -0.0904  0.1147  0.7875  0.0514 -0.2462  0.0252
#>   0.4994 -0.0462  0.6147 -0.5096 -0.4377 -0.2342  0.1336 -0.2717 -0.1177
#>   0.6431  0.1615  0.4110 -0.3444 -0.5291 -0.5048  0.2491 -0.1673 -0.3093
#> 
#> Columns 19 to 20 -0.3162  0.0312
#>   0.3654 -0.2373
#>   0.1445 -0.5630
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4674  0.3443 -0.3869 -0.1168  0.3317 -0.2939 -0.0112 -0.0884 -0.3505
#>  -0.0312 -0.0451  0.1186 -0.2438  0.3925  0.5408  0.3952  0.1232  0.4056
#>  -0.2098 -0.0361 -0.1542  0.3709  0.4810  0.4313 -0.0436  0.0104  0.0633
#> 
#> Columns 10 to 18  0.3897 -0.1111 -0.2175 -0.4146  0.3745 -0.0230  0.2120 -0.5507 -0.1546
#>   0.2666  0.0106  0.2590  0.2962 -0.0692 -0.2617  0.5588 -0.0799  0.2978
#>  -0.0330 -0.1570  0.4287  0.4545  0.0094 -0.1912 -0.1308 -0.1881  0.3220
#> 
#> Columns 19 to 20  0.3909  0.5204
#>  -0.0043  0.5559
#>  -0.1258  0.7206
#> [ CPUFloatType{2,3,20} ]
#>