RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.4320  0.4165  0.6636 -0.1289 -0.3911 -0.3718  0.5697  0.4823 -0.4340
#>  -0.5986  0.5565 -0.2854  0.2332  0.0938 -0.5567 -0.2241  0.8065  0.5463
#>  -0.7536 -0.0711 -0.1513  0.1682 -0.2882  0.6133 -0.3152  0.0065 -0.2107
#> 
#> Columns 10 to 18  0.5270  0.1364  0.3722 -0.3341  0.4040 -0.0658 -0.4152 -0.0772  0.3816
#>   0.1545  0.4532 -0.6642  0.5267 -0.0994  0.1214  0.8010  0.1307 -0.2925
#>  -0.6014  0.5331  0.0674  0.7253 -0.5601  0.4038 -0.6099  0.6491  0.3097
#> 
#> Columns 19 to 20  0.3043 -0.4821
#>   0.6696  0.4668
#>   0.6913 -0.2456
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3567  0.2640  0.1943  0.4406 -0.3956  0.4951 -0.2255  0.9272  0.2073
#>   0.0494  0.5607 -0.1630 -0.1940  0.4695 -0.1471  0.0008 -0.2213 -0.0419
#>   0.2337  0.5833  0.6003  0.0632 -0.2098 -0.2086  0.4477  0.4391 -0.6306
#> 
#> Columns 10 to 18 -0.5574 -0.2566  0.0639 -0.1064  0.5554 -0.2928 -0.0578 -0.5794  0.0281
#>   0.4616  0.4276 -0.2943 -0.1399 -0.3260  0.1476  0.4116 -0.6357  0.3547
#>   0.6708  0.4243 -0.3438 -0.6239  0.0286 -0.0926  0.1604 -0.0372 -0.1086
#> 
#> Columns 19 to 20  0.6440  0.3346
#>   0.0793 -0.2350
#>   0.4882  0.0538
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4094  0.1149  0.4302  0.3014  0.2833  0.5798  0.3017  0.2437 -0.3639
#>   0.3922  0.4597  0.0624 -0.3246 -0.0513 -0.1801  0.1817  0.5935 -0.1767
#>   0.6777  0.5230  0.1006 -0.2581 -0.2168  0.4227  0.2088  0.7018 -0.0259
#> 
#> Columns 10 to 18 -0.2627 -0.3204 -0.6983  0.3955  0.1514 -0.0580  0.0124 -0.7825 -0.0398
#>   0.0322  0.5407 -0.4787  0.0910 -0.1645  0.0063  0.5250 -0.7204 -0.1008
#>  -0.4725 -0.3203 -0.1121  0.1146  0.3151 -0.0584  0.4743 -0.7675 -0.4791
#> 
#> Columns 19 to 20  0.5468 -0.1850
#>   0.5993  0.5834
#>   0.4560  0.2361
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.5589  0.0615  0.3311  0.3837 -0.0806  0.3733  0.3979  0.0625 -0.1768
#>   0.4166  0.1582  0.1711  0.0686  0.4554  0.1900 -0.3714  0.2094  0.1027
#>   0.1476  0.2785  0.5434  0.7199 -0.1509  0.2812  0.1232  0.2773  0.0192
#> 
#> Columns 10 to 18 -0.3631 -0.3034 -0.4939  0.2098  0.2812 -0.2194 -0.0309 -0.4504  0.0228
#>  -0.4239 -0.0201 -0.3259  0.2707  0.0153  0.0396  0.2442 -0.7852  0.2354
#>  -0.6315 -0.2941 -0.0608  0.5441  0.0456  0.0619  0.0443 -0.2640  0.3184
#> 
#> Columns 19 to 20  0.5430  0.3179
#>   0.5289  0.3478
#>   0.8274  0.0285
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.4435  0.0801  0.3168  0.3299 -0.0441  0.1536  0.2895 -0.0933 -0.3514
#>   0.0498 -0.0840  0.3799  0.1654  0.2557 -0.1257 -0.0241 -0.1306  0.0307
#>   0.5071  0.1507  0.2505  0.6094  0.2840  0.0063  0.3953 -0.3701 -0.3139
#> 
#> Columns 10 to 18 -0.2727 -0.3676 -0.3038  0.1726  0.0020 -0.2650 -0.3007 -0.3380  0.0055
#>  -0.1225  0.1405 -0.2453  0.0586 -0.2576 -0.2094 -0.1669 -0.4621  0.2781
#>   0.1704 -0.1022 -0.0987 -0.1238 -0.1207  0.0930 -0.2317 -0.3112  0.3164
#> 
#> Columns 19 to 20  0.5580 -0.5078
#>   0.5777 -0.0574
#>   0.7380 -0.1011
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7878 -0.0783 -0.0921  0.1387 -0.2000  0.2811 -0.0912  0.4835 -0.3558
#>  -0.1298 -0.0340  0.2086 -0.5901  0.3996 -0.6699  0.3503  0.1987 -0.3655
#>  -0.3893 -0.1288  0.5851 -0.1009  0.5255 -0.8440  0.8507  0.3240  0.4334
#> 
#> Columns 10 to 18  0.7879  0.1821  0.4476  0.2037 -0.4657  0.1250  0.1683  0.4850 -0.4832
#>   0.3066 -0.2102  0.6340 -0.1908  0.2569  0.0576 -0.0167 -0.3283 -0.2529
#>   0.3931  0.6115  0.3286  0.4859 -0.5009  0.2636 -0.0013  0.5838 -0.3090
#> 
#> Columns 19 to 20 -0.2410  0.2659
#>   0.6228 -0.2599
#>   0.0730 -0.2167
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4435  0.0801  0.3168  0.3299 -0.0441  0.1536  0.2895 -0.0933 -0.3514
#>   0.0498 -0.0840  0.3799  0.1654  0.2557 -0.1257 -0.0241 -0.1306  0.0307
#>   0.5071  0.1507  0.2505  0.6094  0.2840  0.0063  0.3953 -0.3701 -0.3139
#> 
#> Columns 10 to 18 -0.2727 -0.3676 -0.3038  0.1726  0.0020 -0.2650 -0.3007 -0.3380  0.0055
#>  -0.1225  0.1405 -0.2453  0.0586 -0.2576 -0.2094 -0.1669 -0.4621  0.2781
#>   0.1704 -0.1022 -0.0987 -0.1238 -0.1207  0.0930 -0.2317 -0.3112  0.3164
#> 
#> Columns 19 to 20  0.5580 -0.5078
#>   0.5777 -0.0574
#>   0.7380 -0.1011
#> [ CPUFloatType{2,3,20} ]
#>