RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3291  0.0354 -0.2598  0.8431  0.7834  0.3387  0.0551  0.2621 -0.7092
#>   0.3242 -0.2113 -0.2008  0.4090  0.6286  0.1731 -0.4722  0.0671 -0.1990
#>  -0.9194  0.6926 -0.4968 -0.8200  0.1684 -0.5474 -0.8960  0.2688 -0.1577
#> 
#> Columns 10 to 18 -0.0386  0.8520  0.4245 -0.9385  0.0250  0.2086  0.7067 -0.1196  0.2792
#>  -0.7407  0.9008  0.3558 -0.5219 -0.4587  0.5341  0.3224  0.6216 -0.1672
#>  -0.0819 -0.7410 -0.1238  0.1955  0.6501  0.5083  0.8888 -0.0762 -0.6771
#> 
#> Columns 19 to 20 -0.1703  0.4421
#>  -0.3623  0.8143
#>   0.4267  0.2040
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2190 -0.2002  0.4684  0.0105  0.0180  0.7616 -0.1979  0.3523 -0.2501
#>   0.3383 -0.1420 -0.0223 -0.0518 -0.1331  0.8290 -0.1252  0.4851 -0.3914
#>   0.0049  0.4212 -0.5533 -0.5531  0.2911 -0.4932 -0.3031  0.1505  0.1337
#> 
#> Columns 10 to 18 -0.1468 -0.5274 -0.1348 -0.3996  0.2715 -0.2572  0.2280  0.0169  0.0554
#>  -0.3523 -0.3696 -0.3156 -0.6276  0.1021 -0.1033  0.5597 -0.2813 -0.6325
#>  -0.5993  0.3224 -0.0515 -0.3400 -0.3299 -0.6851  0.5140 -0.1208 -0.1265
#> 
#> Columns 19 to 20  0.0481  0.3398
#>   0.4233  0.1534
#>   0.1948  0.0297
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4440  0.2848  0.0607 -0.7916 -0.2049  0.2003  0.0618  0.1575  0.1622
#>   0.6010 -0.2516  0.2532 -0.2526 -0.6651  0.4567 -0.2704  0.1299  0.1465
#>  -0.1816  0.4470 -0.4186 -0.4517  0.1423  0.1112 -0.1794  0.7499 -0.0104
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4299  0.1059  0.6173  0.1922  0.8445  0.5746 -0.4345 -0.2359  0.1926
#>  -0.4740 -0.0418 -0.2763 -0.0265  0.3129 -0.4677  0.0617  0.6980  0.6446
#>   0.0677  0.0587 -0.3306  0.5141 -0.2011 -0.4920 -0.5227  0.5931 -0.5521
#> 
#> Columns 10 to 18 -0.2753 -0.4058 -0.1881  0.0220 -0.7159  0.0663  0.1797  0.6339  0.8002
#>   0.3206 -0.3358  0.6179  0.2434 -0.2507 -0.1625  0.6183 -0.6325  0.3658
#>   0.1983 -0.4419 -0.0234 -0.4734 -0.1928 -0.3847  0.6961 -0.1536 -0.2830
#> 
#> Columns 19 to 20  0.1416  0.1152
#>   0.5561 -0.9094
#>   0.1845  0.2671
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4437  0.1076  0.1004 -0.5873  0.1318  0.1612 -0.1051  0.1259  0.5590
#>  -0.2394  0.2216 -0.5981 -0.3104 -0.4820 -0.4117 -0.5320  0.4245  0.4612
#>  -0.0629  0.3994 -0.6788 -0.7217  0.2686 -0.6918 -0.3025  0.4839  0.2199
#> 
#> Columns 10 to 18 -0.2121 -0.3920 -0.2526 -0.7547 -0.0386 -0.5484  0.1218 -0.2507 -0.2767
#>   0.5472 -0.6584 -0.1241 -0.1755  0.3283 -0.0491  0.4293 -0.2965 -0.1578
#>   0.1894 -0.6001 -0.0384 -0.2213  0.1285 -0.2932  0.5831 -0.4349 -0.4391
#> 
#> Columns 19 to 20 -0.3284  0.5956
#>   0.1413  0.3920
#>   0.1867 -0.2888
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>