RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5836  0.8783 -0.7982  0.4383 -0.0780 -0.3341 -0.7795 -0.6408 -0.3892
#>  -0.5764 -0.1231  0.3288 -0.1573 -0.7302  0.4687  0.0071 -0.0249  0.1961
#>  -0.6818  0.5084  0.7814  0.4494 -0.3435  0.8639  0.3525  0.7217  0.3281
#> 
#> Columns 10 to 18  0.3350  0.2873  0.1552  0.3344 -0.7100 -0.5049 -0.3584 -0.6633 -0.2556
#>   0.8334  0.7441  0.1799  0.3791 -0.7890  0.4647 -0.4489  0.3126  0.2933
#>   0.6953  0.9160  0.1538 -0.8615 -0.6177  0.8336 -0.5505 -0.7869 -0.5008
#> 
#> Columns 19 to 20  0.9455  0.2369
#>   0.4804 -0.3778
#>  -0.0139  0.4227
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2795  0.2110 -0.1557  0.2208  0.1876 -0.1070 -0.3095 -0.0083  0.0512
#>   0.7161  0.3352  0.4941 -0.0655 -0.1903  0.0020 -0.4587  0.1734 -0.5724
#>  -0.1348  0.4229 -0.0246 -0.1204 -0.4951 -0.7181 -0.8126 -0.1783  0.5637
#> 
#> Columns 10 to 18 -0.3727 -0.1887 -0.0300  0.7508  0.3204 -0.0676  0.3188 -0.2486  0.5294
#>   0.2268  0.1754 -0.2220  0.6896  0.4069 -0.2313 -0.0995 -0.7102  0.1681
#>   0.4098  0.2294  0.5582  0.2544  0.0492  0.5386 -0.2770  0.2732  0.4555
#> 
#> Columns 19 to 20 -0.7357  0.1135
#>  -0.5704  0.5250
#>  -0.1212 -0.0418
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0463  0.5296  0.0903  0.4790 -0.2937 -0.4204 -0.2096 -0.0554 -0.1878
#>  -0.2809  0.2030 -0.0453 -0.2829  0.1768 -0.5000  0.1785 -0.4645 -0.0066
#>   0.0763  0.3534 -0.0701  0.3647 -0.2461 -0.3994 -0.5668 -0.0620  0.0788
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5518 -0.0775 -0.0683  0.5375 -0.3552 -0.1976  0.0833 -0.0456 -0.0710
#>   0.1942 -0.5838  0.0077  0.3023 -0.4187 -0.5794 -0.0939  0.0048 -0.1214
#>  -0.3628 -0.5035 -0.3979  0.2941 -0.6367 -0.0812 -0.8036 -0.2125  0.1920
#> 
#> Columns 10 to 18  0.3228  0.2527 -0.0707 -0.2839  0.3368 -0.5402  0.3357  0.1842 -0.1082
#>   0.0347  0.3072 -0.4952 -0.2236  0.6407  0.6947  0.6425 -0.5958  0.1870
#>  -0.0494  0.7034 -0.6641  0.0634 -0.7481  0.2234  0.5860 -0.7807  0.6837
#> 
#> Columns 19 to 20  0.2825  0.0866
#>  -0.8381  0.1280
#>  -0.7262 -0.0503
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1459  0.3547  0.0079  0.3908 -0.1393 -0.4007 -0.3666  0.0772  0.2946
#>  -0.2613  0.2335 -0.3003  0.0704  0.3011 -0.4164 -0.1090 -0.3919  0.3019
#>  -0.4025  0.2792 -0.3953  0.5277  0.3980 -0.6307 -0.6200 -0.3548  0.2732
#> 
#> Columns 10 to 18  0.3714  0.4210 -0.2769 -0.2401 -0.2557  0.1858 -0.0954 -0.1033  0.3341
#>   0.0885 -0.1627 -0.1458  0.3391  0.0455  0.0603 -0.2742 -0.2927  0.0030
#>  -0.0616 -0.2287  0.1013  0.7167 -0.2610 -0.0974  0.1875 -0.3309  0.2294
#> 
#> Columns 19 to 20  0.0088 -0.1535
#>  -0.4150  0.0439
#>  -0.2107  0.1337
#> [ CPUFloatType{2,3,20} ]
#>