RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0513 -0.6908  0.1772 -0.5006  0.4053 -0.4698 -0.2482  0.3394 -0.3694
#>  -0.9079  0.3313 -0.7719 -0.2360  0.7219 -0.1037  0.3449 -0.4053  0.0820
#>  -0.0615  0.8195  0.7432 -0.1789 -0.8487 -0.9386 -0.1393 -0.4024  0.3010
#> 
#> Columns 10 to 18 -0.6825  0.1376 -0.5874  0.4918  0.4323 -0.3073  0.5766  0.6397 -0.2188
#>   0.9129  0.0284  0.5811  0.9514 -0.4664 -0.6726 -0.7976  0.9025  0.4669
#>   0.2640  0.0201 -0.6431  0.1556 -0.2828 -0.1786 -0.6288 -0.8818  0.0636
#> 
#> Columns 19 to 20 -0.2999  0.0121
#>   0.9648 -0.7432
#>  -0.3220 -0.8718
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4171 -0.2395  0.2098  0.0942 -0.3558 -0.6627 -0.7014 -0.1215  0.2585
#>  -0.2691  0.1620  0.2411 -0.0405 -0.1673 -0.0230 -0.6550 -0.1476 -0.0847
#>  -0.3848 -0.5654 -0.5224  0.1688  0.3323  0.0744  0.3774 -0.3820  0.2715
#> 
#> Columns 10 to 18 -0.1690 -0.1691 -0.0811 -0.2996  0.3288  0.5520 -0.2474  0.4141 -0.6457
#>  -0.3891  0.0191 -0.4456 -0.1344  0.4022  0.2437  0.7028  0.3475  0.5029
#>  -0.3443 -0.4507  0.6499  0.5732 -0.1007  0.0688  0.1485  0.4383 -0.5441
#> 
#> Columns 19 to 20 -0.0169 -0.3710
#>   0.4232  0.2047
#>   0.7947  0.3356
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2751 -0.7784 -0.3282  0.3963  0.3476  0.1119 -0.4458 -0.5428  0.0643
#>   0.0197 -0.3955  0.1071 -0.1938  0.2057 -0.5823 -0.5214 -0.4539  0.2788
#>  -0.4415 -0.3364  0.0243 -0.0949  0.2214 -0.3217 -0.6859 -0.3239 -0.2413
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5725  0.0251 -0.1952 -0.1914  0.3172  0.3435 -0.1376  0.2141 -0.3004
#>  -0.7328  0.1933 -0.4143 -0.5804 -0.1113  0.0071  0.6458  0.2057  0.6859
#>  -0.2959  0.0281 -0.2229 -0.3700 -0.1680  0.6709 -0.3898  0.6628  0.4094
#> 
#> Columns 10 to 18  0.0614  0.0312 -0.1521  0.3288  0.2997 -0.0606 -0.2900 -0.2966  0.1964
#>  -0.7231  0.8897 -0.2982  0.4657  0.2625  0.1985  0.3443 -0.5986 -0.1775
#>  -0.4392 -0.7924 -0.2613 -0.0619  0.4264 -0.4475 -0.5267  0.4813  0.4547
#> 
#> Columns 19 to 20 -0.0287  0.1073
#>   0.1960 -0.5444
#>   0.1209  0.5408
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1487 -0.4853 -0.3076 -0.0070  0.1613 -0.4206 -0.0292 -0.3600  0.2166
#>  -0.5635 -0.4743  0.1683  0.3992 -0.0877 -0.3091 -0.4135 -0.4434  0.2398
#>   0.3937 -0.3297  0.0485 -0.2202 -0.0303 -0.1733 -0.5623  0.1089 -0.1469
#> 
#> Columns 10 to 18 -0.2227 -0.0227 -0.2176 -0.1743  0.0904  0.4419  0.1883  0.5759 -0.1779
#>  -0.1899  0.3283  0.2812  0.2601 -0.3123  0.5909 -0.2861  0.1863  0.1786
#>  -0.5825  0.5477 -0.5109 -0.3261  0.3196  0.6591  0.6260  0.6985 -0.1016
#> 
#> Columns 19 to 20  0.1245  0.0559
#>   0.4881  0.0632
#>  -0.1498  0.2611
#> [ CPUFloatType{2,3,20} ]
#>