RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6255 -0.3534 -0.3887  0.1183  0.7861  0.7336  0.0128  0.0122  0.0228
#>  -0.6650 -0.7824  0.7488 -0.7480  0.4335  0.0119 -0.8344  0.4385  0.0119
#>  -0.0558 -0.8351 -0.0287 -0.8076 -0.3829 -0.2037  0.5171  0.5561 -0.4520
#> 
#> Columns 10 to 18 -0.2897  0.1045  0.5793  0.4571 -0.3383  0.6078  0.9077  0.1363 -0.5541
#>  -0.3346 -0.5018  0.0923  0.7713  0.6639 -0.8997  0.0116  0.0740  0.2979
#>   0.0774  0.9419 -0.4383  0.7657  0.2551  0.4531  0.4481 -0.6618  0.1422
#> 
#> Columns 19 to 20 -0.3525  0.0378
#>  -0.6928 -0.9382
#>   0.3963  0.0289
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0464  0.2807  0.6447 -0.2713 -0.3861  0.4428 -0.1974 -0.6757 -0.0952
#>  -0.0897 -0.3263  0.1482 -0.5441  0.6075  0.7381 -0.2176 -0.2132 -0.5729
#>   0.1176 -0.4380  0.5292 -0.3827 -0.2722  0.5677 -0.6606 -0.2357 -0.6620
#> 
#> Columns 10 to 18 -0.0376 -0.2063  0.4269  0.5640 -0.1610  0.2282 -0.3253 -0.3868 -0.4856
#>   0.1191  0.1441  0.0686  0.4845  0.0589  0.0327  0.3647 -0.2114 -0.5165
#>  -0.3901  0.5267  0.1656  0.1990  0.3600  0.1484 -0.5705  0.2858 -0.0938
#> 
#> Columns 19 to 20 -0.3307  0.4562
#>  -0.1275 -0.4762
#>   0.0366 -0.2993
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1021 -0.0818 -0.2903 -0.6090  0.0224  0.5832  0.2324  0.3173 -0.1907
#>   0.0792 -0.1073  0.6270 -0.3141 -0.5539  0.4933 -0.6228 -0.5298 -0.7430
#>   0.3299 -0.0034  0.2001 -0.2325 -0.5471  0.6073 -0.4385 -0.0585 -0.8031
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7052  0.4724 -0.1864  0.2592  0.4256 -0.7564  0.2557  0.1291 -0.4442
#>  -0.4614 -0.4850  0.2307  0.0307 -0.0517 -0.3635  0.2900  0.6458 -0.0001
#>   0.6072  0.6999 -0.4575  0.8322 -0.2851 -0.0411  0.7285  0.2870 -0.3095
#> 
#> Columns 10 to 18  0.4759 -0.3507 -0.1150  0.4516  0.3942  0.6649  0.9037  0.5834 -0.1390
#>   0.8393 -0.8496 -0.3174  0.2492  0.4955  0.6563  0.0907  0.0359  0.0101
#>  -0.2764 -0.7238 -0.1051  0.2341 -0.3996  0.5618  0.5439  0.1744  0.2882
#> 
#> Columns 19 to 20 -0.2756  0.3780
#>   0.0361  0.2397
#>  -0.4970  0.6134
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1970  0.2024  0.2067 -0.3112 -0.6678  0.6556 -0.4559 -0.2529 -0.5939
#>   0.1894  0.0496  0.2900 -0.2147 -0.6066  0.5374 -0.2745 -0.0257 -0.5380
#>  -0.0083  0.0723  0.3287 -0.0293 -0.5798  0.6462 -0.1611 -0.1521 -0.5346
#> 
#> Columns 10 to 18 -0.2784  0.0530  0.5914 -0.0328 -0.3225  0.6784  0.5386  0.2163  0.1183
#>  -0.1650 -0.4421  0.2856 -0.0917  0.1193  0.4177  0.1363  0.4316  0.1902
#>  -0.0981  0.1555  0.7613 -0.6089 -0.2034  0.6423 -0.0356  0.2312 -0.0306
#> 
#> Columns 19 to 20 -0.4956 -0.0129
#>  -0.2878  0.5883
#>  -0.2502  0.1473
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>