RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5094 -0.2233  0.1492 -0.4374  0.7646 -0.3965  0.4489  0.5886 -0.3679
#>  -0.7717 -0.4004 -0.1641 -0.3093  0.8035  0.4783 -0.0820 -0.3354  0.1083
#>  -0.5225  0.8598  0.2697  0.6030  0.8085 -0.1298  0.5147 -0.9504 -0.9070
#> 
#> Columns 10 to 18 -0.0356 -0.1595 -0.5823 -0.2514 -0.3852  0.2445 -0.2933 -0.3762  0.4439
#>   0.7842  0.6590  0.3174 -0.1447 -0.4309  0.6303 -0.7573 -0.9274  0.1258
#>   0.7381  0.2586 -0.0726 -0.4509  0.5734  0.7491 -0.5428 -0.4024 -0.3314
#> 
#> Columns 19 to 20  0.6437 -0.6030
#>  -0.4944  0.0096
#>  -0.3063  0.0277
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5997 -0.1132  0.2694  0.0323  0.5879  0.3612  0.0668 -0.5051  0.1401
#>   0.1130 -0.2062 -0.1666 -0.6005  0.4354 -0.2098 -0.0637 -0.6640 -0.1088
#>  -0.2291 -0.1839 -0.0433 -0.0779  0.0278 -0.8318  0.1265 -0.1434 -0.5578
#> 
#> Columns 10 to 18  0.0191 -0.2214  0.5774  0.2971 -0.2678  0.1622 -0.6083 -0.2116  0.0207
#>   0.3901  0.4021  0.6413  0.2013 -0.3142 -0.1550 -0.9168  0.0260  0.3702
#>  -0.1730 -0.4509 -0.3180  0.0334 -0.0952 -0.5455 -0.3384 -0.1739 -0.3596
#> 
#> Columns 19 to 20 -0.1099  0.2968
#>   0.1582  0.0797
#>   0.5285 -0.1327
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1212 -0.1367 -0.5046 -0.2416  0.1659 -0.3548  0.1142 -0.1151 -0.2108
#>   0.1837  0.1668 -0.3675  0.3158  0.5341 -0.2494 -0.0608 -0.3884 -0.4026
#>  -0.4694  0.0693  0.1117  0.2484  0.5987  0.2017  0.0343  0.2782  0.2575
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7362  0.1803 -0.8178  0.5919  0.1832  0.4041 -0.4255  0.3186  0.4272
#>   0.2801  0.4673 -0.4757  0.1502  0.2888 -0.0414 -0.3311  0.2752 -0.0710
#>  -0.6980  0.5514 -0.6066  0.9171 -0.7674  0.1715 -0.2293 -0.4547  0.3181
#> 
#> Columns 10 to 18  0.0993  0.1756  0.3425 -0.6818  0.5237 -0.3833 -0.4519  0.5105  0.1955
#>   0.0509 -0.5703  0.5352 -0.2245 -0.4299 -0.5061  0.2662 -0.0964 -0.2618
#>   0.2438  0.0100  0.3785  0.5651  0.7477  0.5287 -0.2425  0.2020  0.2159
#> 
#> Columns 19 to 20 -0.3693  0.0246
#>  -0.0074 -0.0568
#>  -0.4787  0.8244
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2543  0.3594 -0.6533 -0.0013  0.1197 -0.6120 -0.1387 -0.5842 -0.2770
#>  -0.2504  0.1902 -0.5357 -0.3306  0.5651 -0.0193  0.1600 -0.4526 -0.1149
#>  -0.1540  0.0103 -0.5424  0.0586  0.6426 -0.6271 -0.1138 -0.3250 -0.6513
#> 
#> Columns 10 to 18  0.4355 -0.1072  0.3143  0.4809  0.0230 -0.4147 -0.6586 -0.0406  0.0590
#>   0.2676  0.6337  0.3631  0.0946  0.0005 -0.0833 -0.5576 -0.4634  0.1026
#>   0.6203  0.1943 -0.4115  0.2252  0.1465 -0.0929 -0.4652 -0.5259 -0.0381
#> 
#> Columns 19 to 20  0.3359  0.0609
#>  -0.0472  0.0834
#>   0.4179 -0.0038
#> [ CPUFloatType{2,3,20} ]
#>