RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1266 -0.6257  0.1541 -0.3432  0.6680 -0.1074  0.5424  0.5337 -0.5946
#>   0.2398 -0.5051  0.4462  0.8138  0.6816 -0.2971 -0.6786 -0.3540 -0.2229
#>  -0.2152 -0.6299 -0.9536 -0.5072  0.2876 -0.2858 -0.2875 -0.9102  0.2260
#> 
#> Columns 10 to 18 -0.5845 -0.5478 -0.4289 -0.1799  0.2416 -0.5160  0.7350 -0.0482 -0.7370
#>   0.1421  0.4609 -0.4986 -0.4547 -0.7940 -0.2544 -0.4786  0.1062  0.7811
#>   0.9659  0.0627 -0.6977  0.4242  0.8336  0.0513 -0.0445 -0.4557 -0.3783
#> 
#> Columns 19 to 20  0.1902  0.3253
#>   0.4516  0.5429
#>  -0.0869  0.5675
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0661 -0.4151  0.0089 -0.4449  0.3215  0.2425  0.2895 -0.0888 -0.1108
#>  -0.7027 -0.5577 -0.5229 -0.6896  0.7216  0.5245  0.1358  0.1357  0.1726
#>  -0.7315 -0.0501  0.2470  0.1063 -0.4122 -0.0359 -0.8468  0.0438  0.4562
#> 
#> Columns 10 to 18 -0.5096 -0.1156 -0.5552 -0.5206 -0.0063 -0.1983 -0.2334 -0.2384 -0.0445
#>   0.3618 -0.1801 -0.1368  0.0793 -0.2469 -0.3038  0.0399  0.2404 -0.6909
#>  -0.2461  0.0708 -0.6019  0.5124 -0.2887  0.2738 -0.7113 -0.0285  0.2805
#> 
#> Columns 19 to 20 -0.1509  0.3887
#>  -0.0459 -0.1707
#>  -0.2447 -0.4212
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2166 -0.2997  0.0953 -0.2342  0.1614 -0.2892 -0.6931  0.1274  0.1221
#>  -0.2636 -0.2371  0.2218  0.3463  0.0726  0.1322 -0.4336  0.1889 -0.2867
#>  -0.5491 -0.2826  0.1390 -0.5893  0.4734  0.3147  0.2053  0.0932 -0.0999
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7870  0.1007  0.5299  0.3325 -0.1837  0.0811  0.2006 -0.1299  0.6828
#>  -0.0419  0.4837  0.1540  0.2776  0.0074  0.1979  0.3405  0.0556 -0.1675
#>  -0.1082  0.1245 -0.1648  0.3853 -0.7039  0.0715  0.0157 -0.1393 -0.2168
#> 
#> Columns 10 to 18 -0.1937 -0.5122  0.1865 -0.1641 -0.7034 -0.0721  0.0752 -0.4201 -0.6806
#>   0.0107  0.5130  0.4442  0.5434  0.1802  0.2613 -0.0104 -0.1391  0.2568
#>   0.7838  0.7240  0.5377 -0.1946  0.8994 -0.3658 -0.6664 -0.2419  0.3759
#> 
#> Columns 19 to 20  0.1114 -0.7183
#>  -0.1225 -0.1766
#>   0.0289 -0.1376
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6724  0.1927 -0.2165  0.2181 -0.2835  0.2729 -0.4925  0.0678 -0.1054
#>  -0.3635 -0.1949 -0.2390 -0.3681  0.2630  0.2813  0.0042  0.0335 -0.3405
#>  -0.3083 -0.1602  0.6085 -0.8302  0.3586  0.1963  0.1415  0.5234  0.0176
#> 
#> Columns 10 to 18  0.0265 -0.2284 -0.4310  0.2792 -0.2015  0.2314 -0.2205  0.2032  0.3744
#>  -0.1916  0.4050 -0.6356 -0.1025 -0.0864  0.1398 -0.3845 -0.3616  0.1357
#>  -0.0030 -0.4140 -0.2508  0.3820  0.2183 -0.2352  0.2095  0.0646 -0.6260
#> 
#> Columns 19 to 20 -0.1091 -0.0795
#>  -0.3695 -0.2152
#>   0.1649  0.0048
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>