RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7981 -0.5980  0.4813 -0.1193 -0.9814  0.4840 -0.0089  0.9259 -0.4255
#>   0.7715  0.1575 -0.3274  0.8306 -0.3761 -0.7822 -0.7153 -0.0229  0.3913
#>  -0.7375 -0.9265  0.7567 -0.0876 -0.9651  0.7241  0.5941  0.8552  0.0857
#> 
#> Columns 10 to 18 -0.4260 -0.4321  0.5972  0.2411  0.0463 -0.3376 -0.1909  0.8340 -0.3612
#>  -0.5193 -0.0259  0.8604 -0.7583 -0.5237 -0.2827 -0.6349  0.7299  0.8927
#>   0.2285  0.8573 -0.1327 -0.3282 -0.2178 -0.5333 -0.6019 -0.5777 -0.5279
#> 
#> Columns 19 to 20  0.7913 -0.4369
#>   0.8178 -0.6193
#>   0.8074 -0.8347
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4109 -0.5174 -0.0278  0.3540 -0.3652 -0.5249  0.1892 -0.0855 -0.7978
#>   0.6053 -0.7073  0.3743  0.5008 -0.2209 -0.6939 -0.1071  0.2778 -0.6012
#>   0.4844  0.0001 -0.3036  0.0317 -0.7910 -0.3000 -0.0355  0.2253 -0.2586
#> 
#> Columns 10 to 18 -0.3938  0.3284  0.0623  0.5362  0.6827  0.1749 -0.1053 -0.2742 -0.3390
#>  -0.5820  0.3373  0.4693  0.1953  0.2684 -0.3679 -0.1670 -0.3860  0.2382
#>  -0.2729 -0.3233  0.3922 -0.0332  0.2684  0.4408 -0.0249  0.1254 -0.5478
#> 
#> Columns 19 to 20  0.0244  0.0607
#>   0.1289 -0.5443
#>   0.7430  0.5207
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2707 -0.3014  0.0094  0.2705 -0.6391 -0.2223  0.2504  0.7927  0.0877
#>   0.0084 -0.3752  0.0735  0.0884 -0.5136 -0.4871 -0.4298  0.4426 -0.3930
#>   0.1032 -0.4230  0.2392  0.4290 -0.6211 -0.5643  0.3072  0.6296  0.0789
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2975 -0.1098 -0.3175  0.6566  0.4676 -0.3835 -0.1218  0.1996  0.1501
#>   0.7691  0.1421 -0.6057  0.2018 -0.3691 -0.3290 -0.2074 -0.4880 -0.1835
#>   0.3509 -0.0183 -0.2371  0.0188 -0.0916 -0.0029  0.5607  0.2331 -0.3483
#> 
#> Columns 10 to 18 -0.5482 -0.1956  0.0545 -0.2979 -0.3592 -0.3197  0.4858  0.4357 -0.3039
#>  -0.3551 -0.3688  0.0069  0.2442 -0.0966 -0.1716  0.4007 -0.4471  0.4055
#>   0.6009 -0.2125  0.9163 -0.5397 -0.3371  0.4131 -0.6295 -0.5582 -0.0625
#> 
#> Columns 19 to 20 -0.0123 -0.2549
#>  -0.8253 -0.1354
#>   0.6319  0.1051
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0191 -0.4490  0.0300  0.1232 -0.5069 -0.5852 -0.1870  0.3331  0.0568
#>   0.1968 -0.6136  0.0903  0.2413 -0.6572 -0.5534 -0.0856 -0.0556 -0.0970
#>   0.5307 -0.5777  0.4426  0.1768 -0.5395 -0.4539  0.1846  0.7747 -0.1859
#> 
#> Columns 10 to 18 -0.6191  0.3390  0.7322 -0.3107 -0.0254  0.5678 -0.0993 -0.0666  0.0504
#>  -0.1783  0.0508  0.6179 -0.3450 -0.0059  0.1017 -0.6220 -0.1869  0.0723
#>  -0.6768  0.6771  0.6125 -0.2258  0.8179  0.3953  0.4832 -0.0289 -0.2699
#> 
#> Columns 19 to 20  0.5074 -0.2862
#>   0.2833 -0.3892
#>   0.1816  0.0267
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>