RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6248 -0.5821  0.2824 -0.8406  0.4586  0.9020 -0.1873 -0.5800 -0.7122
#>   0.1677  0.7980 -0.8490  0.6281 -0.7843 -0.4181  0.3522  0.6206  0.8680
#>   0.0547  0.1297  0.1461  0.4442 -0.0100  0.5579 -0.1956 -0.4445 -0.1123
#> 
#> Columns 10 to 18  0.4666 -0.2080  0.7097 -0.8000 -0.5077  0.4371  0.0090 -0.6997 -0.7817
#>   0.2870 -0.1692 -0.5863  0.1721 -0.3548 -0.6046 -0.1903  0.7404  0.5578
#>  -0.0242 -0.0350  0.9402 -0.0630 -0.1174  0.2590 -0.3220 -0.5416  0.7930
#> 
#> Columns 19 to 20  0.4539 -0.0781
#>   0.3723 -0.1794
#>  -0.5222  0.0337
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0373  0.4072  0.4173 -0.1236  0.0096 -0.4605 -0.0503 -0.4522  0.5895
#>  -0.2764 -0.0702  0.4298  0.1679 -0.2692  0.5109  0.3079 -0.2556 -0.3789
#>  -0.1079  0.3558  0.1569 -0.0534  0.1646  0.2811  0.4050  0.0681 -0.1762
#> 
#> Columns 10 to 18 -0.8152 -0.1333 -0.1127  0.0455 -0.0506  0.1142  0.0532  0.0382 -0.1563
#>   0.0204 -0.4770  0.0341  0.1814  0.2604  0.2682 -0.6530 -0.2737  0.3829
#>   0.3067 -0.2646 -0.1026  0.0630  0.0541  0.5234 -0.0940 -0.7014 -0.6771
#> 
#> Columns 19 to 20  0.3103  0.5518
#>  -0.5406 -0.0303
#>   0.8396 -0.0579
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0303  0.2533 -0.0503 -0.0479 -0.3136 -0.1416  0.4426 -0.3257 -0.1946
#>  -0.1862  0.4439  0.0565 -0.0213 -0.0532 -0.1949 -0.1844  0.2128 -0.0674
#>  -0.2739 -0.0306  0.2028 -0.0679 -0.1860 -0.3285 -0.0524 -0.7217  0.5985
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1050  0.2669 -0.0361  0.0408 -0.2431  0.4958  0.3246  0.2206  0.3079
#>   0.4145 -0.2632  0.0421  0.1436  0.5890 -0.2011 -0.1161  0.3200  0.0182
#>   0.4817 -0.3877  0.1306 -0.2545 -0.0061 -0.3737  0.8187 -0.5356  0.5825
#> 
#> Columns 10 to 18 -0.1087 -0.8784  0.8278 -0.2479  0.2423  0.3076  0.2001 -0.2521  0.1729
#>  -0.2564  0.3253  0.0587 -0.2656  0.4783 -0.7171 -0.4176 -0.3769 -0.4741
#>  -0.2023 -0.2081  0.7248  0.0149 -0.3199 -0.0614 -0.4841 -0.2156  0.6604
#> 
#> Columns 19 to 20  0.7492 -0.4237
#>   0.0554 -0.1555
#>  -0.0810 -0.6928
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2184 -0.0353  0.1250 -0.1309 -0.4354 -0.0879  0.4398 -0.5102  0.0862
#>   0.1052  0.3759  0.1980  0.2350 -0.6203  0.0140 -0.2989  0.0174  0.1943
#>   0.1254  0.1262  0.0297  0.0603 -0.3664  0.2918  0.7108 -0.3261 -0.0921
#> 
#> Columns 10 to 18 -0.4508 -0.3534 -0.0576 -0.5236 -0.1709  0.3918 -0.0638 -0.2913  0.1384
#>   0.2397 -0.5017  0.1882 -0.1399  0.1359 -0.2620  0.0070  0.0175 -0.0244
#>  -0.6289 -0.2326  0.0588 -0.5506 -0.1075  0.4058 -0.6687 -0.4704 -0.2357
#> 
#> Columns 19 to 20  0.1027  0.4851
#>   0.1458 -0.0162
#>  -0.1537  0.1382
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>