RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4999  0.5322 -0.1481  0.9374  0.5257  0.8586  0.0232 -0.5345  0.9138
#>   0.8909  0.7556  0.0081  0.2510 -0.4885 -0.1715 -0.0709 -0.6179 -0.4027
#>  -0.4983 -0.4356 -0.8488 -0.2831  0.6965 -0.2993  0.4844 -0.0598 -0.0586
#> 
#> Columns 10 to 18 -0.0012  0.6727 -0.0014  0.1336 -0.8795 -0.6438  0.7367  0.7157  0.7704
#>   0.5579  0.0026  0.2322 -0.5997 -0.9425  0.6176 -0.7711 -0.0269  0.2458
#>   0.0755  0.3738 -0.8126 -0.0184 -0.2292  0.2640 -0.9314  0.1056 -0.1486
#> 
#> Columns 19 to 20  0.0103  0.0840
#>  -0.0179  0.3671
#>   0.1611  0.7954
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4883  0.3863 -0.0089 -0.1513  0.1070  0.2133 -0.0337  0.1390  0.5800
#>  -0.7810  0.0640  0.1869  0.5214 -0.3652 -0.2980 -0.3108 -0.0682 -0.2292
#>  -0.5666  0.3264  0.0638  0.7434 -0.0361  0.3712 -0.1588 -0.0424  0.0866
#> 
#> Columns 10 to 18  0.3597 -0.4852  0.0334 -0.1711 -0.0398 -0.6289 -0.0725 -0.1606  0.0998
#>   0.1526 -0.4302 -0.6249 -0.0393 -0.1255  0.0964  0.3304  0.5203  0.5568
#>   0.4100  0.2038  0.0942 -0.6957  0.0178 -0.2398 -0.1940  0.3317  0.3719
#> 
#> Columns 19 to 20 -0.0130  0.4390
#>  -0.2055 -0.2999
#>  -0.3553  0.2879
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0512  0.0463  0.5926  0.1376  0.2416 -0.4336  0.1679  0.3539  0.2870
#>   0.1691  0.2997  0.2117  0.2092 -0.2556  0.0278 -0.6262  0.0807 -0.1021
#>  -0.6406  0.2397  0.4994  0.4455 -0.5231 -0.1118 -0.2653 -0.1617 -0.2750
#> 
#> Columns 10 to 18  0.3944  0.0835  0.0237 -0.3170 -0.2276  0.1897 -0.2053 -0.4137  0.2469
#>   0.5912 -0.0522  0.1913 -0.1854 -0.0776 -0.0070  0.2278 -0.1751  0.6270
#>   0.6944 -0.5280 -0.1340 -0.3720  0.4449 -0.0746 -0.2524  0.2573  0.3054
#> 
#> Columns 19 to 20 -0.0276  0.2794
#>  -0.1559  0.0605
#>   0.0332  0.3368
#> 
#> (4,.,.) = 
#>  Columns 1 to 9 -0.0922  0.4578  0.2669  0.3543 -0.2161 -0.3543 -0.2276  0.3418 -0.3062
#>  -0.0375  0.1723  0.0280  0.0960 -0.1087 -0.4846 -0.3559  0.1707 -0.3154
#>  -0.0468  0.4328  0.2638  0.2269 -0.4208 -0.2115 -0.7487 -0.0007 -0.3356
#> 
#> Columns 10 to 18  0.8225 -0.1960 -0.2733 -0.5337  0.0470  0.2458 -0.2733 -0.0564 -0.1222
#>   0.7432  0.1693 -0.3977 -0.4323  0.0803  0.3395 -0.5374 -0.1626  0.0047
#>   0.8134  0.2749 -0.2223 -0.2683  0.3225  0.3648 -0.4629 -0.1790  0.0134
#> 
#> Columns 19 to 20 -0.5041  0.2786
#>  -0.2695  0.4649
#>   0.0338  0.1043
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.3381  0.4285  0.2866  0.7053 -0.5669 -0.2453 -0.3101  0.1711 -0.3615
#>  -0.6654  0.0485  0.1742  0.5719 -0.1044 -0.4195 -0.1309 -0.1187 -0.3975
#>  -0.4697 -0.1448  0.2194  0.5516 -0.7715 -0.3361 -0.3543 -0.4224 -0.2021
#> 
#> Columns 10 to 18  0.4936  0.3595 -0.1928 -0.5603  0.2382  0.3291  0.0754  0.2571  0.4855
#>   0.5355 -0.0019 -0.4430 -0.6027  0.3726  0.3411 -0.0815 -0.1671  0.6975
#>   0.5456  0.0587 -0.5003 -0.2807  0.4464  0.2645 -0.4766  0.2770  0.5223
#> 
#> Columns 19 to 20  0.0565 -0.2962
#>  -0.4395  0.1496
#>  -0.1156 -0.2440
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1136 -0.6790  0.4435  0.4478 -0.2515  0.3819 -0.2761 -0.3323  0.1744
#>  -0.5607 -0.6585 -0.6991 -0.2159 -0.5622  0.4070 -0.4969 -0.0159 -0.2109
#>  -0.0141 -0.0892  0.3271  0.2383 -0.0896 -0.6885 -0.4866  0.1707 -0.3215
#> 
#> Columns 10 to 18  0.0832 -0.2674  0.1314  0.2087 -0.5903  0.4102 -0.2944 -0.4505  0.2319
#>  -0.8863 -0.0344 -0.1306 -0.3456  0.6951 -0.2858 -0.1821 -0.2562 -0.3198
#>  -0.7551  0.1167 -0.3533  0.8262  0.4040 -0.6954 -0.3130  0.3843 -0.0915
#> 
#> Columns 19 to 20 -0.0981  0.0950
#>  -0.3508 -0.8232
#>   0.2168 -0.3532
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3381  0.4285  0.2866  0.7053 -0.5669 -0.2453 -0.3101  0.1711 -0.3615
#>  -0.6654  0.0485  0.1742  0.5719 -0.1044 -0.4195 -0.1309 -0.1187 -0.3975
#>  -0.4697 -0.1448  0.2194  0.5516 -0.7715 -0.3361 -0.3543 -0.4224 -0.2021
#> 
#> Columns 10 to 18  0.4936  0.3595 -0.1928 -0.5603  0.2382  0.3291  0.0754  0.2571  0.4855
#>   0.5355 -0.0019 -0.4430 -0.6027  0.3726  0.3411 -0.0815 -0.1671  0.6975
#>   0.5456  0.0587 -0.5003 -0.2807  0.4464  0.2645 -0.4766  0.2770  0.5223
#> 
#> Columns 19 to 20  0.0565 -0.2962
#>  -0.4395  0.1496
#>  -0.1156 -0.2440
#> [ CPUFloatType{2,3,20} ]
#>