RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1182  0.9342 -0.7030  0.5269  0.3113 -0.7918  0.4497 -0.2666  0.7546
#>   0.4047  0.6216  0.5617  0.2078  0.7731 -0.1922  0.9064 -0.5406  0.7171
#>  -0.7270  0.6864 -0.2065 -0.2274  0.0573 -0.4618  0.1234 -0.3296  0.1619
#> 
#> Columns 10 to 18 -0.2562 -0.1043  0.3110 -0.6767  0.4469 -0.0616 -0.3405  0.7711  0.1667
#>  -0.3129 -0.3708 -0.5034 -0.1321  0.3508  0.6694 -0.0914 -0.1378 -0.2352
#>  -0.5047  0.3462 -0.0347  0.1777 -0.2418 -0.3338  0.7604  0.5780  0.2364
#> 
#> Columns 19 to 20  0.7985 -0.0842
#>   0.5146 -0.3707
#>   0.3959  0.5808
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7241  0.5952 -0.3993  0.6005  0.6055 -0.2088 -0.6681  0.0407 -0.5741
#>  -0.2497  0.1331  0.2685  0.5136  0.1441 -0.6346 -0.3748  0.2406  0.4649
#>  -0.4829  0.6783  0.0065 -0.1935  0.2884  0.0729  0.0369  0.4023 -0.0040
#> 
#> Columns 10 to 18  0.8429  0.7686 -0.3798 -0.5744 -0.3929 -0.0133 -0.7646 -0.7017 -0.1085
#>   0.3071  0.0344  0.3839  0.2250 -0.1564 -0.1581 -0.0255 -0.2237 -0.1061
#>   0.2987  0.5086 -0.1444 -0.4131 -0.0361 -0.0632 -0.6264  0.1057  0.1618
#> 
#> Columns 19 to 20 -0.3663 -0.3805
#>   0.4699 -0.5598
#>   0.3087  0.3029
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3647  0.2010  0.2213  0.2268  0.0956  0.2099  0.0367  0.2762 -0.7220
#>   0.1054  0.4596 -0.2167  0.7221 -0.0585 -0.4009 -0.1209 -0.1829 -0.6470
#>  -0.3264  0.2107  0.2203  0.0233 -0.2501  0.1245  0.3484  0.3182 -0.3885
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6345  0.0618  0.0944  0.4083 -0.2530 -0.8189  0.1516  0.4212  0.6281
#>  -0.1259 -0.2334 -0.4931  0.0710 -0.6173 -0.3464  0.2060 -0.2663 -0.3122
#>  -0.4784  0.1256 -0.5313 -0.1583  0.3000 -0.5039  0.4972  0.3830 -0.3985
#> 
#> Columns 10 to 18  0.2076  0.3791  0.5490 -0.4248  0.1819 -0.1893 -0.3641  0.6290 -0.8399
#>   0.3136 -0.0518  0.4519  0.0851  0.6298  0.4180 -0.6683 -0.4259  0.0141
#>  -0.6373 -0.5510  0.5396  0.3315 -0.5988 -0.2045  0.5672  0.1359 -0.0453
#> 
#> Columns 19 to 20  0.5524 -0.5254
#>   0.0471 -0.5176
#>   0.4035 -0.4628
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0601  0.4923 -0.4792  0.1294  0.0959 -0.6383 -0.2497  0.1926  0.0296
#>  -0.1319  0.0217  0.1284  0.1666 -0.2601 -0.2274  0.6502 -0.1074 -0.0061
#>   0.0631  0.0723 -0.1137  0.7106  0.0158 -0.7004 -0.3757 -0.1233 -0.1834
#> 
#> Columns 10 to 18 -0.2580  0.4737 -0.0425  0.3011 -0.3636  0.2509  0.2075  0.4601 -0.0008
#>  -0.2771  0.4419 -0.1569 -0.1963 -0.3750  0.3318 -0.3531  0.2029 -0.1295
#>   0.4163  0.4822 -0.3242 -0.1169 -0.2487  0.1023 -0.2997 -0.5161 -0.5175
#> 
#> Columns 19 to 20  0.4041  0.1956
#>  -0.2814  0.3844
#>  -0.2354 -0.5157
#> [ CPUFloatType{2,3,20} ]
#>