RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6447 -0.0157 -0.4115  0.3778  0.3742  0.3731  0.1367  0.2669  0.2050
#>  -0.0527 -0.5784  0.2696  0.0029 -0.6808  0.0647 -0.8379 -0.2291 -0.1875
#>  -0.6281  0.6742 -0.8253 -0.4605  0.3372 -0.3782 -0.0619  0.8494  0.6028
#> 
#> Columns 10 to 18  0.5942  0.6461  0.3912  0.1218  0.3847  0.2746  0.6048 -0.1655 -0.0533
#>  -0.5255  0.8072 -0.3507 -0.9067 -0.8468 -0.0881  0.2356  0.3628  0.5450
#>   0.3574  0.2471 -0.4897  0.8492  0.9142  0.7202 -0.4973  0.7657 -0.5686
#> 
#> Columns 19 to 20  0.5164  0.2189
#>   0.7832  0.4292
#>  -0.4685  0.9306
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2027 -0.3869  0.2151  0.2614 -0.6317 -0.2044  0.0056 -0.2529  0.0346
#>  -0.6837  0.0122 -0.6213  0.3797 -0.2427 -0.4481 -0.5547 -0.2533  0.7000
#>   0.4446 -0.0417  0.1447 -0.3030 -0.0368  0.3854  0.6772 -0.0201 -0.2076
#> 
#> Columns 10 to 18 -0.1111  0.6062 -0.0330 -0.3027 -0.4051 -0.1498  0.1591  0.0046  0.0413
#>  -0.3543 -0.1873 -0.3959  0.1388 -0.1862  0.0338 -0.6399  0.5470 -0.2783
#>   0.4758  0.4243  0.1995  0.4930  0.0618  0.4633  0.8314 -0.5535 -0.4216
#> 
#> Columns 19 to 20  0.2233  0.1301
#>  -0.4033  0.8641
#>   0.6888 -0.5720
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3810  0.0960 -0.7042  0.2654 -0.1438  0.1419 -0.1402  0.1840  0.6906
#>   0.0461 -0.0652 -0.1499 -0.1843  0.5429 -0.0356  0.0460  0.1355  0.3047
#>   0.0443  0.0649 -0.3077  0.2420 -0.5873 -0.1664  0.0576 -0.0838  0.2156
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1710 -0.2510  0.0100  0.3317 -0.2333  0.2133  0.1541 -0.4084 -0.0788
#>  -0.8202  0.4083  0.4519  0.3742 -0.3131  0.3117 -0.2336  0.1077 -0.5661
#>   0.3664  0.1556  0.0700  0.7389  0.1183  0.3080  0.2711 -0.2367  0.4171
#> 
#> Columns 10 to 18 -0.5129  0.4969 -0.1524  0.0476 -0.3256 -0.5546  0.1451 -0.6054  0.5043
#>  -0.2380  0.1521  0.2900 -0.5133  0.6125 -0.2974 -0.4512 -0.0965  0.6076
#>   0.1459  0.6428  0.1057 -0.1837 -0.7437 -0.1472  0.6235 -0.2799  0.4435
#> 
#> Columns 19 to 20  0.1332 -0.0520
#>  -0.6709 -0.2038
#>   0.3636 -0.7742
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3441  0.6258  0.0636  0.0243 -0.0495 -0.1079  0.4010  0.1417  0.0237
#>  -0.0096 -0.1237 -0.4272  0.1447  0.4560 -0.1701  0.2450  0.0041  0.4656
#>  -0.2195  0.5629  0.2649  0.0636  0.2120  0.3867  0.4493  0.2055  0.1945
#> 
#> Columns 10 to 18  0.2481 -0.0361 -0.1579  0.1530  0.0165  0.0265 -0.2592  0.2091 -0.3410
#>   0.3686  0.2328 -0.1770  0.7530  0.1355  0.0573  0.2890  0.0854  0.1991
#>   0.4325  0.1721  0.2308 -0.1012  0.2048  0.0010 -0.1442  0.0105 -0.3085
#> 
#> Columns 19 to 20  0.4085 -0.4237
#>  -0.0015  0.1904
#>   0.3822 -0.4340
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>