RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6660 -0.7111  0.6723  0.3255 -0.6046 -0.8981 -0.3166 -0.1039  0.6495
#>  -0.0197 -0.3063 -0.5235  0.0153 -0.5832 -0.4950 -0.8021  0.5580  0.3386
#>   0.0029  0.0991  0.0377 -0.3238 -0.0282 -0.7852  0.0118  0.5960  0.4402
#> 
#> Columns 10 to 18  0.5960 -0.5308  0.8031  0.1039 -0.5915  0.6814  0.5006  0.3088  0.5371
#>   0.0626 -0.1204  0.6409 -0.7934 -0.3338  0.3301 -0.4441  0.6636 -0.7082
#>  -0.7724  0.7918  0.0161 -0.2480 -0.1166 -0.1928  0.2642 -0.6141 -0.0643
#> 
#> Columns 19 to 20  0.8568  0.9087
#>   0.7448  0.0147
#>  -0.0558  0.7074
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4603 -0.5273  0.0692 -0.2067 -0.3578  0.0042 -0.2419  0.4578 -0.5134
#>  -0.6195 -0.5952  0.0235 -0.1782 -0.1565  0.2187 -0.4085  0.3637 -0.1278
#>   0.1627 -0.6826 -0.1230  0.1975 -0.2724 -0.3222 -0.2793  0.7251 -0.3559
#> 
#> Columns 10 to 18 -0.3464  0.6238  0.1250 -0.2171 -0.1617 -0.0444 -0.1282 -0.2124 -0.2550
#>  -0.8384  0.8080 -0.2453 -0.1362 -0.7939 -0.2066  0.5170  0.0571  0.0415
#>  -0.6406  0.3576  0.3269  0.0292 -0.3120  0.3155  0.4816  0.2144  0.1552
#> 
#> Columns 19 to 20 -0.1461  0.4059
#>  -0.6618  0.5878
#>   0.0303  0.4339
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2476 -0.2484  0.0378  0.0708 -0.3618 -0.2797 -0.1257  0.5418 -0.4627
#>   0.2752  0.1598  0.1034  0.1720 -0.0009 -0.1060 -0.1899  0.7920 -0.0834
#>  -0.5905 -0.2855 -0.2396 -0.2642 -0.2634  0.0207 -0.4225 -0.1810  0.2866
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1298  0.0761 -0.0559  0.7147 -0.2146  0.0959  0.0810 -0.3105 -0.6150
#>   0.2938 -0.3573 -0.6916 -0.1628  0.0987  0.6453  0.0011 -0.1221 -0.6491
#>  -0.2123 -0.4903 -0.2073 -0.4539 -0.4073  0.5213 -0.2902  0.8248 -0.4296
#> 
#> Columns 10 to 18  0.3459  0.1230  0.6544  0.0573 -0.5577  0.5398  0.5970 -0.5188  0.4909
#>  -0.1530  0.0863  0.1577  0.6267 -0.0752  0.0497  0.2729 -0.4015  0.5984
#>  -0.5564 -0.8115 -0.6837  0.1983 -0.6090 -0.6653 -0.8291 -0.3133  0.1996
#> 
#> Columns 19 to 20 -0.3710 -0.4152
#>   0.4563 -0.9282
#>  -0.7662 -0.4677
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4313 -0.1505 -0.0406  0.1688  0.2420  0.1915 -0.4644  0.4601 -0.4536
#>   0.2295 -0.0274  0.1517  0.1425  0.0875 -0.0965 -0.2504  0.7972 -0.1193
#>  -0.2012  0.3594  0.0573  0.1432  0.1470 -0.2163 -0.1439  0.4851 -0.3334
#> 
#> Columns 10 to 18 -0.1007  0.4562 -0.0135 -0.6275 -0.2421  0.2600  0.3747  0.0004  0.2040
#>  -0.4996  0.3917  0.4005 -0.1184 -0.2001  0.4491  0.4000 -0.2419  0.2072
#>  -0.2404 -0.0348  0.1464 -0.1501  0.0046  0.6300  0.3786  0.3059 -0.1143
#> 
#> Columns 19 to 20 -0.4771  0.4501
#>   0.1960  0.5911
#>   0.7592  0.6181
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>