RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7004 -0.7780 -0.3538 -0.0287 -0.6803 -0.7466  0.7449  0.4910  0.7638
#>   0.3048  0.0070  0.7737  0.0373  0.0467 -0.4799  0.3078  0.7232  0.4134
#>   0.8781  0.6098  0.7496  0.6630 -0.4291  0.1651 -0.1519  0.8672  0.5607
#> 
#> Columns 10 to 18  0.5273 -0.6290  0.0357  0.2746 -0.8331 -0.7139 -0.2207  0.8045  0.8246
#>   0.0137 -0.0058  0.3982 -0.5369 -0.0274  0.6609  0.0356  0.0884 -0.0885
#>   0.8115 -0.3429 -0.0135 -0.1427 -0.1021  0.5994 -0.2961  0.7332 -0.4417
#> 
#> Columns 19 to 20  0.3629 -0.2227
#>   0.1547  0.0726
#>   0.3543 -0.2283
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6523  0.2704  0.7942  0.4858 -0.3883 -0.6247  0.5940  0.1129  0.0392
#>   0.2500  0.3139  0.4755  0.4839 -0.2193  0.3279  0.1658 -0.1847 -0.7981
#>  -0.0107  0.5553 -0.3803  0.2068 -0.3484  0.7354  0.0553 -0.4371 -0.6728
#> 
#> Columns 10 to 18 -0.1532 -0.6558  0.6435 -0.2871  0.6369  0.6210 -0.0552 -0.3486  0.4516
#>   0.0877  0.6088  0.6244  0.4724  0.3666  0.1867  0.0509 -0.1779  0.3611
#>  -0.0627  0.3831  0.3403 -0.0664  0.2282 -0.0267  0.0962 -0.1394  0.2940
#> 
#> Columns 19 to 20  0.1522  0.6119
#>  -0.5651 -0.3091
#>  -0.6566 -0.2747
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1232  0.0445  0.5639  0.2917 -0.0363 -0.0260  0.6226  0.3022 -0.0888
#>   0.3019  0.3375 -0.2640 -0.3173  0.1508 -0.0311 -0.1955 -0.1985 -0.4405
#>  -0.0015  0.2402 -0.1830 -0.5195 -0.1968  0.0067  0.3983 -0.0378 -0.2208
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5435  0.6624  0.1229 -0.2532  0.2295  0.0555 -0.5299  0.1770 -0.1535
#>  -0.1971  0.7405  0.7632 -0.3323  0.3542 -0.0334 -0.7070 -0.5930 -0.5745
#>  -0.3485  0.3912  0.0756 -0.4321  0.0096 -0.7772  0.3185 -0.0115 -0.2539
#> 
#> Columns 10 to 18  0.1447  0.0692 -0.0141  0.3796  0.4409  0.2017 -0.1384  0.4026 -0.6221
#>  -0.1125  0.8210  0.0204 -0.0479  0.2328  0.4527  0.3977 -0.3972 -0.1720
#>  -0.4893 -0.1564  0.2605 -0.0451  0.0277  0.3973  0.1400 -0.1217 -0.0596
#> 
#> Columns 19 to 20  0.3871 -0.1546
#>   0.8664 -0.2476
#>   0.0528 -0.6042
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2625  0.3330 -0.4179 -0.0757 -0.2100 -0.2359  0.2307  0.0604  0.1509
#>   0.3695  0.5742  0.2220 -0.4521 -0.5549  0.0352  0.6782  0.1411 -0.2126
#>  -0.1272  0.5832  0.3548 -0.1793 -0.5310  0.1189  0.1487  0.0894 -0.6898
#> 
#> Columns 10 to 18 -0.2034  0.3313 -0.1739 -0.2957  0.0452  0.1853 -0.3171  0.2515  0.0117
#>   0.4336  0.0089  0.0536 -0.1903 -0.1278  0.3599 -0.3194  0.1969 -0.1837
#>   0.3352  0.3327  0.0046 -0.1387 -0.1681  0.4797 -0.4508  0.1251  0.1440
#> 
#> Columns 19 to 20 -0.4756  0.2616
#>  -0.5877  0.3563
#>  -0.4433  0.0802
#> [ CPUFloatType{2,3,20} ]
#>