RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1607 -0.3110  0.7354  0.0903 -0.3694 -0.5684 -0.6984 -0.1373 -0.7831
#>   0.0526 -0.0037 -0.3379 -0.2729  0.2888 -0.2634 -0.7332  0.5052  0.9201
#>   0.6063 -0.5254  0.5617  0.6539 -0.4415  0.5932 -0.6253  0.5228  0.6150
#> 
#> Columns 10 to 18 -0.4179  0.2348  0.4453 -0.7638 -0.2205  0.3822 -0.6100  0.6134  0.7310
#>   0.1245  0.1461  0.1721  0.1994 -0.1238 -0.2418 -0.0342  0.0279  0.4100
#>  -0.5771 -0.6877  0.7239 -0.7511 -0.8600  0.6491 -0.1183 -0.0161 -0.5486
#> 
#> Columns 19 to 20 -0.5136 -0.1022
#>   0.5444 -0.4951
#>   0.9152  0.7736
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2222 -0.1199  0.3125 -0.5385  0.1810  0.5619  0.8241  0.6158  0.0457
#>   0.0451  0.2672 -0.1191 -0.2860  0.5378  0.1163 -0.4817  0.1538  0.3117
#>   0.0736 -0.3082 -0.0003 -0.0761  0.6479  0.2232 -0.3350 -0.1067  0.7563
#> 
#> Columns 10 to 18 -0.3520  0.0952 -0.2294 -0.7921  0.0359  0.3555 -0.4103 -0.4224  0.7002
#>   0.2113 -0.0330 -0.3025 -0.0005  0.0463 -0.4853  0.3650 -0.0154  0.3677
#>   0.3519 -0.3076  0.1816 -0.0314 -0.0887 -0.6666 -0.2646 -0.3294  0.4656
#> 
#> Columns 19 to 20  0.1878 -0.2359
#>  -0.2560 -0.5655
#>   0.0298  0.0028
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2368  0.1237 -0.2779  0.2293  0.1779  0.4992  0.1027 -0.4305 -0.4179
#>  -0.2277 -0.3636 -0.1327 -0.0783 -0.2410 -0.1426 -0.3261  0.1343  0.2370
#>   0.2927  0.4539  0.1195 -0.0814  0.7010  0.7010 -0.2359 -0.1400 -0.1390
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0986 -0.0492 -0.4994 -0.2150  0.3000  0.2844 -0.4458 -0.4311 -0.0807
#>  -0.4276  0.2345 -0.2461  0.0187 -0.8199  0.5764  0.2582 -0.5116  0.1201
#>   0.2276  0.3195 -0.0782 -0.1181 -0.0143 -0.1996  0.2938 -0.7635  0.0065
#> 
#> Columns 10 to 18 -0.3609 -0.0882 -0.2291  0.1027 -0.0848  0.7499  0.4849  0.3294 -0.3004
#>  -0.2413  0.4029  0.2216 -0.1182  0.5790  0.4200  0.9118  0.5748 -0.9037
#>   0.3307  0.2701 -0.1282 -0.5513  0.2583  0.4879 -0.1734  0.4907  0.4023
#> 
#> Columns 19 to 20  0.3582 -0.6078
#>  -0.3079 -0.7483
#>  -0.2707 -0.2228
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3815 -0.0282  0.3626 -0.2528  0.5070  0.6380  0.1339 -0.0290 -0.0416
#>   0.4292 -0.2671  0.3931  0.2981  0.0175  0.5601 -0.4074 -0.3330  0.4149
#>   0.3176  0.0152 -0.1487  0.0352  0.2151  0.1936 -0.6059 -0.4096  0.1566
#> 
#> Columns 10 to 18  0.0802  0.1514  0.2594 -0.5704  0.0113  0.2498 -0.1707 -0.3372  0.4835
#>   0.0329 -0.0271  0.2310 -0.2056 -0.0149  0.2007  0.1723  0.1188  0.2773
#>   0.4235 -0.0152  0.2931 -0.1552 -0.3363 -0.3115  0.5199  0.1618  0.2271
#> 
#> Columns 19 to 20  0.0408 -0.0055
#>  -0.0715  0.4727
#>  -0.0644 -0.1869
#> [ CPUFloatType{2,3,20} ]
#>