RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3655  0.3278 -0.0991  0.2727 -0.3600 -0.4163 -0.9084 -0.0174 -0.5535
#>   0.1774 -0.7375  0.4235  0.6636  0.0422  0.6841 -0.3288 -0.1167 -0.7752
#>   0.6131  0.4720 -0.8936  0.3016  0.1753  0.0470 -0.7993  0.6118 -0.8056
#> 
#> Columns 10 to 18 -0.4686 -0.6568 -0.0019  0.5969  0.1435  0.5701  0.2797  0.6898  0.8394
#>   0.8407 -0.4391 -0.7435 -0.6458 -0.5736 -0.1592  0.3152  0.7406  0.4384
#>  -0.7875 -0.1568  0.3747  0.7963  0.3756 -0.7580 -0.2287  0.1953  0.9056
#> 
#> Columns 19 to 20 -0.2412 -0.2455
#>   0.4283 -0.1311
#>  -0.6724 -0.3279
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1489  0.0409 -0.0605  0.2217 -0.5951  0.2856 -0.1397  0.1012 -0.3297
#>   0.3755 -0.5773  0.3449 -0.1178 -0.0571  0.1287 -0.4286  0.2416 -0.6887
#>   0.0956 -0.4065 -0.6538  0.4731 -0.3365 -0.2746 -0.4496  0.4036  0.0450
#> 
#> Columns 10 to 18 -0.4484  0.4517  0.7676 -0.2306  0.2335  0.0017 -0.1664 -0.0740  0.1344
#>   0.2697  0.7399  0.2792 -0.4323  0.0382 -0.4775  0.3049 -0.0891  0.2999
#>  -0.2316  0.8211  0.3329 -0.5072 -0.7010  0.6419  0.0257 -0.3387 -0.2464
#> 
#> Columns 19 to 20 -0.1773 -0.5724
#>   0.3815 -0.5089
#>  -0.0907  0.2073
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1123 -0.4161 -0.5399  0.3545 -0.3780 -0.1544  0.2013  0.1348 -0.5280
#>   0.2135 -0.4731 -0.4068  0.2115 -0.1006 -0.1927 -0.1594  0.4253 -0.4788
#>   0.1737 -0.8107  0.0754 -0.3726 -0.5922  0.0036 -0.5336 -0.0718 -0.4360
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7487  0.0959 -0.6994 -0.3662  0.4462 -0.5852  0.2213  0.7393  0.5656
#>  -0.2633  0.0188  0.1686  0.5513 -0.1452 -0.4129 -0.4335  0.4201  0.5344
#>  -0.8131  0.0753  0.5550  0.4674 -0.4384  0.7807 -0.5811  0.7454  0.0854
#> 
#> Columns 10 to 18 -0.2419 -0.2240 -0.0542  0.2601  0.4252  0.4163  0.3085 -0.6460  0.5475
#>  -0.6075 -0.6609  0.4494  0.4387  0.2285  0.5168 -0.2103 -0.7232  0.3871
#>  -0.3462 -0.4154 -0.2122 -0.6927  0.2203  0.4586 -0.1485 -0.5659  0.2903
#> 
#> Columns 19 to 20  0.1478  0.0961
#>   0.5527 -0.2115
#>   0.2913  0.6544
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5658 -0.7331 -0.1998  0.1111 -0.3168  0.1480 -0.6170  0.2413 -0.5868
#>   0.0979 -0.5978 -0.2252 -0.0163 -0.1609  0.3360 -0.7008  0.6313 -0.4382
#>   0.3181 -0.3648 -0.4626  0.1764  0.0287  0.2438 -0.1573  0.6074 -0.2597
#> 
#> Columns 10 to 18  0.1778  0.2820 -0.2508 -0.4043 -0.1965 -0.0605 -0.2402  0.0207 -0.2865
#>   0.2792  0.4696 -0.2192 -0.7709 -0.4049 -0.0439 -0.0648  0.3694 -0.1346
#>   0.1788  0.3318 -0.0544 -0.6673 -0.1082 -0.3839 -0.2407  0.2697 -0.1206
#> 
#> Columns 19 to 20  0.0923  0.0046
#>  -0.4883 -0.1126
#>  -0.3207 -0.1014
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>