RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2405  0.5726 -0.7044 -0.5593 -0.0065  0.2354  0.4908  0.0981 -0.1302
#>  -0.8625  0.6066 -0.4166 -0.7257 -0.8680  0.9303  0.2816 -0.7443  0.2710
#>   0.7395 -0.2774 -0.6988 -0.7351 -0.7353  0.2238  0.7878 -0.6280  0.3934
#> 
#> Columns 10 to 18  0.6397 -0.2215 -0.4968  0.7923  0.6748 -0.3840  0.0303  0.3179  0.0535
#>  -0.0304  0.4971 -0.3418  0.3406  0.0165 -0.8511  0.8982 -0.0864  0.0674
#>  -0.0349  0.5216 -0.7195  0.8091  0.4555 -0.4676  0.5072  0.3282  0.4668
#> 
#> Columns 19 to 20 -0.2595 -0.7062
#>  -0.5044 -0.0984
#>  -0.8578 -0.9123
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0367  0.4749 -0.2035 -0.0802  0.3957  0.0656 -0.2228 -0.0477 -0.1291
#>  -0.4163  0.0679  0.2096 -0.6072 -0.2387  0.6457  0.4438 -0.1937  0.4008
#>   0.3040 -0.0669 -0.2275 -0.0450  0.5773  0.1709  0.3841 -0.3894 -0.0000
#> 
#> Columns 10 to 18  0.0158  0.1198 -0.3837 -0.3508 -0.5313  0.4286 -0.0709  0.5157  0.0766
#>   0.2631  0.6281 -0.0691 -0.2960 -0.5785  0.5913  0.3248  0.0501 -0.0729
#>  -0.0926  0.3626  0.0994  0.3451 -0.3243  0.8480 -0.2859 -0.4197 -0.2100
#> 
#> Columns 19 to 20 -0.3928 -0.5692
#>  -0.5970 -0.3936
#>  -0.3662 -0.3596
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0052  0.2427 -0.3528  0.1169 -0.0948  0.2104 -0.3044  0.1752  0.0710
#>  -0.4905 -0.2385  0.0771 -0.3610  0.2143  0.5424 -0.1051 -0.1373 -0.0065
#>  -0.0758 -0.6410  0.4287 -0.0341  0.1760  0.2946 -0.1743  0.0074  0.0008
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1874  0.3504  0.3753  0.0963 -0.1363 -0.2059 -0.1591 -0.0531 -0.1111
#>   0.0428 -0.0256 -0.0102 -0.0793 -0.0599  0.3526 -0.3808 -0.3353 -0.0837
#>  -0.0890 -0.3121 -0.2332 -0.0863 -0.7357 -0.0637 -0.5253 -0.4223  0.2496
#> 
#> Columns 10 to 18 -0.1601  0.1970  0.1770 -0.5354 -0.1310 -0.1623 -0.1532  0.1635 -0.2993
#>  -0.0767  0.3124 -0.5226  0.2078 -0.0364  0.1392  0.1958  0.5194  0.3462
#>  -0.7100 -0.3001 -0.7246  0.2087  0.0280  0.0161  0.5733  0.4668 -0.1487
#> 
#> Columns 19 to 20  0.0614  0.0933
#>  -0.7520 -0.5244
#>  -0.6225 -0.2223
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1865  0.1694 -0.3329  0.0300  0.3438 -0.0919 -0.1787 -0.0247 -0.0404
#>  -0.2568 -0.0150 -0.0908 -0.5345  0.0357  0.0713 -0.1166 -0.2954 -0.3663
#>  -0.2339 -0.4714  0.3222 -0.5558 -0.0245  0.1398 -0.2487 -0.0300  0.0086
#> 
#> Columns 10 to 18  0.1760  0.1675 -0.4701 -0.2712 -0.5457  0.0779  0.0094 -0.1137 -0.5022
#>   0.0118  0.4083  0.0715 -0.0132  0.2574  0.0748  0.2000  0.0613 -0.2568
#>  -0.1509  0.3845 -0.2536 -0.0067 -0.2277  0.4839  0.2318  0.0299 -0.6474
#> 
#> Columns 19 to 20 -0.3414 -0.4284
#>  -0.1201 -0.4013
#>  -0.6531 -0.4922
#> [ CPUFloatType{2,3,20} ]
#>