RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8057 -0.2394 -0.4414  0.4284 -0.0308 -0.2020  0.8530 -0.2192 -0.6409
#>   0.4726  0.0106 -0.6611  0.3576  0.1874  0.4017 -0.9217  0.4331  0.5076
#>  -0.1711 -0.1157  0.3458 -0.8694  0.9318 -0.3120 -0.5575  0.5561  0.7503
#> 
#> Columns 10 to 18 -0.3810  0.4207  0.3072 -0.1921 -0.6232 -0.6927 -0.6186 -0.4210  0.4456
#>  -0.2562 -0.6851  0.3758 -0.5032 -0.3281  0.7618 -0.2665  0.1650  0.5873
#>  -0.0887 -0.5561  0.8232 -0.0908  0.1790  0.4167  0.0504 -0.9409  0.5544
#> 
#> Columns 19 to 20 -0.2298 -0.8807
#>   0.7866 -0.2108
#>   0.8417 -0.3751
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5670  0.0521  0.2599  0.2860  0.3817  0.5969  0.6618 -0.1824  0.8292
#>   0.2399 -0.1484  0.0674  0.5408  0.4222 -0.2438 -0.1655  0.2771  0.5260
#>   0.2723 -0.1561 -0.5330  0.0859  0.3725 -0.0592  0.5293  0.2907  0.6619
#> 
#> Columns 10 to 18  0.7101  0.2437  0.0758 -0.5056  0.7676  0.5078  0.4750  0.2359 -0.2099
#>  -0.6023 -0.4130 -0.0438 -0.1670  0.0397  0.3657 -0.3063  0.3339 -0.5369
#>   0.1517 -0.5449  0.5812 -0.0047 -0.0645 -0.3942 -0.5288  0.3202 -0.2405
#> 
#> Columns 19 to 20 -0.0785 -0.4409
#>   0.3717 -0.0320
#>   0.2861 -0.5597
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3395  0.3067 -0.3413  0.2406  0.5730 -0.6214  0.4771 -0.0937  0.1800
#>  -0.1795  0.0966 -0.6901  0.3298  0.4284 -0.3454 -0.0509  0.3440  0.2327
#>   0.5185  0.0903 -0.4861  0.3040  0.5849 -0.5291 -0.6186  0.6917  0.6809
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2871  0.2869  0.0988  0.5743  0.7380 -0.3615 -0.2760  0.0727  0.3010
#>  -0.1919 -0.0907 -0.4551  0.5275 -0.1389  0.6854 -0.3262  0.9126  0.5641
#>  -0.8383 -0.0016 -0.5496 -0.1404 -0.4416 -0.4719 -0.5934  0.8095 -0.1182
#> 
#> Columns 10 to 18  0.0510 -0.0154  0.6460 -0.2216 -0.2106  0.4251  0.0923 -0.4429 -0.0451
#>   0.1614 -0.5947  0.8295  0.4405  0.4056 -0.3544 -0.2847  0.0902 -0.2987
#>   0.4506  0.0976  0.9515 -0.4164  0.6188  0.0826 -0.1938 -0.0505 -0.4907
#> 
#> Columns 19 to 20 -0.3956 -0.7475
#>  -0.0683  0.6535
#>   0.1173 -0.1240
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4893 -0.0302 -0.2216  0.2358  0.3064 -0.2717  0.2898  0.5383  0.6787
#>   0.2982 -0.2585 -0.5247 -0.0249  0.2620 -0.1681  0.6027 -0.1393  0.4781
#>   0.4093  0.1327 -0.1371  0.4661  0.3625 -0.0081  0.2347  0.3917  0.6539
#> 
#> Columns 10 to 18 -0.1912 -0.3788  0.4660 -0.2682  0.5646 -0.5141 -0.5885  0.0795  0.5104
#>  -0.5393 -0.3155  0.2823 -0.4301  0.1356 -0.3535 -0.3250 -0.0753  0.6913
#>  -0.3084 -0.3158  0.4139 -0.4429  0.6678  0.0802 -0.2448  0.2118  0.3558
#> 
#> Columns 19 to 20  0.2612 -0.5884
#>   0.4758 -0.2567
#>   0.3247 -0.3540
#> [ CPUFloatType{2,3,20} ]
#>