RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5218 -0.2588  0.7975  0.2598 -0.3118  0.4069  0.7614  0.1988 -0.2719
#>  -0.0406 -0.2604 -0.3836  0.2940 -0.7914  0.2840  0.8314  0.5738 -0.4161
#>   0.4744  0.6183 -0.4156  0.2602 -0.3054  0.1087  0.7137  0.7764 -0.4943
#> 
#> Columns 10 to 18 -0.9821 -0.0916 -0.6022 -0.0836 -0.7678  0.6462  0.7001 -0.8214  0.0786
#>  -0.5389  0.3353  0.4550 -0.3597 -0.4160  0.2791  0.9197 -0.1143 -0.9200
#>  -0.8354 -0.8202 -0.8570 -0.4013 -0.8960  0.1337  0.1348 -0.1744 -0.8102
#> 
#> Columns 19 to 20 -0.6112 -0.7207
#>  -0.2295 -0.4965
#>  -0.6188  0.3050
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7557  0.2632 -0.4573 -0.3696  0.6015 -0.6637  0.3354  0.1413  0.2810
#>  -0.4300  0.5077 -0.5150  0.2597  0.3284 -0.1583  0.5046  0.1342 -0.0636
#>  -0.4832 -0.0337 -0.5802 -0.2084  0.2350  0.1495  0.6242  0.6558  0.2529
#> 
#> Columns 10 to 18 -0.3791  0.3297 -0.0318  0.7913 -0.3369  0.2001  0.3330 -0.1052 -0.1913
#>  -0.4905  0.0328  0.0208  0.1483 -0.2176  0.1341  0.1519 -0.0040 -0.4531
#>  -0.4377  0.1845  0.0505  0.1182 -0.7976  0.5557  0.1795 -0.3148 -0.3511
#> 
#> Columns 19 to 20 -0.0886 -0.1642
#>  -0.4536 -0.1967
#>  -0.3193 -0.2738
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3899  0.3029 -0.2688  0.3586  0.0433  0.5968  0.1782 -0.0148  0.3339
#>   0.0073  0.2992 -0.5219 -0.2056  0.2098  0.3529  0.4579  0.4122  0.3094
#>  -0.1078  0.0113 -0.2024  0.3068  0.2591  0.0325  0.0620  0.2328 -0.0396
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1033  0.8287 -0.3295 -0.7572  0.2925 -0.0064  0.0546  0.7703  0.3315
#>  -0.1608 -0.5219  0.3391 -0.4085  0.2398  0.4458 -0.3806 -0.2047  0.3504
#>  -0.8215  0.4340 -0.1714  0.2000  0.7228 -0.2617  0.8233  0.3973  0.6584
#> 
#> Columns 10 to 18  0.1904 -0.2945 -0.0070  0.3373  0.2236 -0.3535 -0.1139 -0.5354  0.2436
#>   0.8472  0.6813  0.7542  0.2705 -0.2935  0.2957  0.3524  0.2567  0.4844
#>  -0.0647  0.0749  0.4391  0.5852  0.3990 -0.2243  0.0825 -0.6435 -0.0561
#> 
#> Columns 19 to 20 -0.1602 -0.6027
#>  -0.6405  0.1960
#>  -0.1528 -0.4467
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1723  0.4387 -0.0625 -0.0842  0.2185 -0.1111  0.0533  0.3838  0.1291
#>  -0.0231  0.2530 -0.3345 -0.0354  0.0482  0.0155  0.5613  0.4818  0.2254
#>   0.0729  0.5089 -0.3504 -0.1687 -0.0605  0.1789  0.5230  0.3646  0.3450
#> 
#> Columns 10 to 18 -0.4099 -0.0257  0.2805  0.4555  0.0480 -0.1171  0.4060 -0.4913 -0.0323
#>  -0.5266  0.1297  0.1364  0.2973  0.3405  0.2821 -0.3518  0.4201 -0.1948
#>  -0.1738  0.0419 -0.0849  0.1930 -0.1604  0.2395  0.7941 -0.4789 -0.2957
#> 
#> Columns 19 to 20  0.1026 -0.3112
#>  -0.0618  0.0904
#>  -0.3518 -0.1405
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>