RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0443  0.3294  0.5682  0.3423  0.0767 -0.1376 -0.0456 -0.0496  0.4457
#>  -0.1995 -0.9017  0.4278 -0.7325  0.0231 -0.1573 -0.5351  0.6217 -0.0408
#>   0.6251 -0.2784 -0.4001 -0.8497 -0.7348 -0.9859  0.1326 -0.7107 -0.4969
#> 
#> Columns 10 to 18 -0.7806  0.9230  0.0843  0.0365 -0.7622 -0.8007 -0.4918 -0.8856 -0.3330
#>  -0.1193 -0.8059 -0.8635 -0.1891  0.3826 -0.1733 -0.6475  0.8120  0.6696
#>   0.0374 -0.5076  0.5572  0.4306 -0.5627 -0.8261 -0.7241  0.3744  0.8784
#> 
#> Columns 19 to 20 -0.0962  0.1771
#>   0.6242 -0.5641
#>  -0.3526  0.6907
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1945  0.1228 -0.0837 -0.4258 -0.0252 -0.6708  0.4508 -0.2873 -0.3475
#>   0.0296  0.1362  0.5711  0.0459 -0.7838 -0.0009  0.4982  0.5095  0.1954
#>   0.6444 -0.4180  0.7077 -0.5206 -0.2974 -0.4313  0.4368 -0.2681  0.5186
#> 
#> Columns 10 to 18 -0.3411  0.1415 -0.2316  0.1922  0.1412 -0.1439  0.0060 -0.2296 -0.2659
#>  -0.7680 -0.1681  0.5236 -0.7027 -0.1796 -0.7151 -0.4699  0.2400  0.2389
#>  -0.4008 -0.2635  0.0894 -0.2414 -0.2092 -0.7482 -0.4307 -0.1490  0.6491
#> 
#> Columns 19 to 20 -0.1586  0.3898
#>  -0.4599 -0.2277
#>  -0.0340  0.4823
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0171 -0.4456 -0.1205 -0.1565 -0.2189 -0.3172  0.3889 -0.3421  0.2203
#>  -0.2858 -0.4678 -0.4739 -0.3769 -0.6145 -0.7461  0.0297 -0.0233 -0.6935
#>   0.1063  0.1528  0.4035 -0.4691 -0.6098 -0.4496  0.3901  0.3028  0.4945
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3448 -0.4313 -0.6841 -0.5296 -0.5019 -0.2146 -0.2760 -0.1666 -0.5699
#>   0.2782 -0.0821  0.0225  0.1497  0.1085  0.4150  0.3565  0.4257 -0.4340
#>   0.0279  0.3678  0.6396 -0.4578 -0.0010 -0.5590  0.3510  0.4538  0.1780
#> 
#> Columns 10 to 18 -0.1438  0.9090  0.3306  0.1945  0.3857  0.6060 -0.8749 -0.0691  0.0110
#>   0.6477  0.3057 -0.0751 -0.3731 -0.6693 -0.5503  0.0397 -0.1211  0.4466
#>   0.4343  0.3687 -0.1443  0.2307  0.0360 -0.1663  0.6726  0.0029  0.0921
#> 
#> Columns 19 to 20  0.8121 -0.2046
#>  -0.1803 -0.4747
#>   0.0724 -0.5535
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0988 -0.4116  0.5278 -0.2679 -0.5225 -0.4902  0.4130  0.1005  0.5335
#>   0.4092 -0.4998  0.4925 -0.1627 -0.3583 -0.3799  0.2458  0.0949  0.3045
#>  -0.1589 -0.2711 -0.1428 -0.4640 -0.1865 -0.5697  0.2881 -0.0340 -0.4216
#> 
#> Columns 10 to 18 -0.4174 -0.6002  0.2443 -0.1236  0.2073 -0.6602  0.1264 -0.1420  0.4285
#>  -0.0183 -0.1754 -0.6421  0.5359  0.0556 -0.3209  0.1783  0.2875  0.3365
#>  -0.2542  0.0802 -0.2267  0.5941 -0.5408 -0.5345 -0.0724  0.4435  0.2470
#> 
#> Columns 19 to 20  0.4308  0.1087
#>  -0.0853  0.3727
#>   0.1177  0.1263
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>