RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0296 -0.7126 -0.2523  0.2058  0.2658  0.8837 -0.4416 -0.0376 -0.2406
#>  -0.0615 -0.1578 -0.6939  0.1956  0.4045  0.2231 -0.0056 -0.3549 -0.4418
#>  -0.0134 -0.9263 -0.7182  0.2539  0.8607  0.4708  0.8147  0.2624 -0.4503
#> 
#> Columns 10 to 18  0.5413  0.5556  0.2625  0.7959 -0.5921 -0.0912 -0.0062  0.0600 -0.5248
#>   0.4615  0.2029 -0.5488 -0.2353 -0.3265  0.5384 -0.6855 -0.9472  0.5831
#>  -0.5930  0.4099  0.3891  0.6150 -0.7170  0.1598 -0.6113 -0.9624  0.6470
#> 
#> Columns 19 to 20 -0.3781 -0.3286
#>   0.5622 -0.3949
#>  -0.1826  0.0834
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0607 -0.3172  0.0529  0.4151  0.4833  0.3502 -0.6619  0.0005 -0.0495
#>  -0.5537 -0.6698 -0.0403 -0.3149  0.6005 -0.0011  0.3179 -0.0678 -0.1343
#>  -0.4081 -0.2979 -0.2987 -0.2543  0.4951  0.6258 -0.1619  0.2204 -0.2897
#> 
#> Columns 10 to 18  0.0441  0.0864  0.1136 -0.1448  0.1613 -0.0527  0.2780 -0.5820  0.1630
#>  -0.1836  0.0722 -0.0182  0.4496 -0.1272  0.1723 -0.1202 -0.3531 -0.0549
#>   0.2039 -0.1720 -0.2905 -0.6327  0.3935  0.0854  0.3728 -0.7822  0.3970
#> 
#> Columns 19 to 20  0.1738 -0.1421
#>   0.0574 -0.1322
#>   0.4111  0.6187
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0049 -0.5009  0.0873  0.2362  0.7180  0.0242 -0.3084  0.1342  0.0653
#>  -0.0880 -0.3777 -0.3497 -0.0248  0.5480 -0.2271  0.0056  0.0070  0.4835
#>  -0.1515 -0.5715 -0.2028  0.1385  0.7746 -0.1042 -0.0764  0.1274  0.4660
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1828 -0.5411  0.2651  0.0061  0.0015  0.2267 -0.5288 -0.6835  0.0554
#>  -0.1025 -0.5895  0.4258  0.1483 -0.5633  0.3016 -0.4811 -0.5105 -0.2494
#>   0.2120 -0.2198  0.1152 -0.5152 -0.1049  0.4897 -0.1970  0.2678 -0.0779
#> 
#> Columns 10 to 18 -0.4110 -0.3522  0.1252 -0.0506 -0.1181  0.1340  0.1570 -0.0122  0.0642
#>  -0.6759 -0.1514  0.6629  0.2232  0.6567  0.7954  0.2306 -0.6859 -0.5576
#>  -0.6691 -0.4422 -0.3424  0.1014  0.1536  0.1135  0.0359  0.0160 -0.0573
#> 
#> Columns 19 to 20 -0.6426  0.1199
#>  -0.7986  0.0749
#>  -0.2409 -0.2329
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1533 -0.5334  0.1591 -0.1741  0.4758 -0.1720 -0.0983 -0.0771 -0.1277
#>  -0.4083 -0.6160  0.0142 -0.2873  0.6078 -0.2142  0.1408 -0.1122 -0.0661
#>  -0.2548 -0.4874 -0.1911 -0.1129  0.6125  0.1459 -0.0219  0.1046  0.0093
#> 
#> Columns 10 to 18 -0.2081  0.1487  0.2372  0.5498  0.1859 -0.0660 -0.1654 -0.6379  0.0652
#>   0.0122  0.1402  0.1866  0.4457 -0.3484 -0.4653 -0.4634 -0.7698  0.0360
#>   0.2076  0.3539 -0.0332  0.3343  0.1134  0.1386 -0.2461 -0.6355  0.3494
#> 
#> Columns 19 to 20  0.1561 -0.1567
#>  -0.2768  0.1841
#>   0.1645  0.1444
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>