RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8113  0.1004 -0.5724 -0.0812 -0.4475 -0.2627  0.8753  0.6871 -0.7526
#>   0.9020 -0.1583 -0.4868  0.2913  0.8413  0.1870  0.4639 -0.0793  0.5753
#>   0.8890 -0.1638  0.8619  0.5062 -0.2688 -0.4016  0.5988  0.5444  0.5892
#> 
#> Columns 10 to 18 -0.6061  0.1975  0.6668 -0.8705  0.2308 -0.6072 -0.6442  0.4743  0.6888
#>   0.2145 -0.3286  0.8692 -0.1685  0.4023 -0.6924 -0.1683  0.1996  0.6548
#>  -0.3056  0.0543 -0.1751 -0.3727 -0.1347  0.9069  0.1850 -0.5202  0.5303
#> 
#> Columns 19 to 20 -0.2990 -0.3760
#>  -0.0217 -0.5022
#>  -0.9297  0.5676
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6222  0.0902 -0.3080 -0.2167 -0.2004  0.2421  0.6134  0.5343  0.4378
#>   0.5071 -0.2470 -0.4956 -0.0836  0.1609 -0.0557  0.2678 -0.0175 -0.1816
#>   0.0096 -0.3108  0.0075  0.5804 -0.1012  0.3840 -0.2692  0.1448  0.6690
#> 
#> Columns 10 to 18 -0.5770  0.0708 -0.2062 -0.1855  0.3235  0.1210  0.0645  0.3303 -0.1050
#>  -0.5504  0.3219  0.3399  0.4932  0.3265 -0.3509 -0.3081 -0.0352  0.0091
#>  -0.4684  0.0429 -0.3236  0.2517  0.3770  0.1369  0.3931  0.3290  0.3526
#> 
#> Columns 19 to 20 -0.4693 -0.1754
#>   0.3546 -0.4499
#>  -0.5148  0.0590
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1826  0.2648 -0.0488 -0.2227 -0.0803  0.2249  0.3531 -0.1018  0.1133
#>   0.4695  0.6164  0.1802 -0.3694  0.2202 -0.2243  0.3591  0.1474 -0.1876
#>  -0.1210  0.0466  0.0460  0.2019 -0.1666  0.1517  0.0043 -0.6528 -0.2272
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5174 -0.4233 -0.0665  0.5015  0.1424 -0.3750 -0.2226 -0.2770  0.0240
#>  -0.6049 -0.3331 -0.6246  0.1073  0.4917  0.1497  0.2467 -0.5174  0.1452
#>  -0.2192 -0.4410 -0.0655  0.2466 -0.0526 -0.3897  0.3665  0.0377 -0.1594
#> 
#> Columns 10 to 18 -0.5187 -0.2481 -0.4009 -0.0978  0.3422  0.0582  0.5262 -0.1337 -0.2276
#>  -0.2112  0.2842 -0.1711  0.1393  0.3805 -0.0689  0.4765  0.1052  0.7313
#>   0.1771 -0.5773 -0.0721 -0.4616  0.5496 -0.1562  0.4496  0.5234 -0.2927
#> 
#> Columns 19 to 20 -0.3038  0.3037
#>  -0.1870 -0.0184
#>   0.5224  0.2654
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6685  0.0942 -0.0260  0.1076 -0.1567  0.0870  0.5737  0.2600  0.1445
#>   0.4676 -0.1199 -0.1663  0.1117 -0.2139  0.0683  0.3419 -0.0208  0.1269
#>   0.5363  0.3242  0.2293  0.1189  0.2405 -0.0796  0.4226  0.2829  0.0916
#> 
#> Columns 10 to 18 -0.2248  0.4649 -0.0701  0.0842  0.2252 -0.1124 -0.2385  0.3047  0.1318
#>   0.0299  0.2334 -0.1823  0.4450  0.3085 -0.1862 -0.5614 -0.0340 -0.4012
#>  -0.4859  0.5217  0.4752 -0.4636  0.1132 -0.1316 -0.3669  0.3223  0.2342
#> 
#> Columns 19 to 20  0.2588 -0.5859
#>  -0.0843 -0.1993
#>  -0.1978 -0.3609
#> [ CPUFloatType{2,3,20} ]
#>