RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1284 -0.7590  0.6808 -0.4434 -0.1948  0.7256  0.1489  0.7275 -0.7761
#>  -0.7538  0.2491 -0.5144 -0.5856  0.8650 -0.1331 -0.5487  0.7452  0.4254
#>   0.7084 -0.0239  0.6513 -0.1310  0.3503  0.2760  0.5986 -0.1001 -0.1244
#> 
#> Columns 10 to 18  0.3110  0.7731  0.3884 -0.8937  0.7876  0.6327 -0.1194 -0.2786 -0.7362
#>  -0.2665 -0.5092 -0.7654 -0.1741 -0.8060  0.4216  0.9733 -0.6875 -0.8944
#>  -0.6224  0.7980  0.1110 -0.3057 -0.0941 -0.1063  0.1111  0.8052 -0.2186
#> 
#> Columns 19 to 20 -0.1233 -0.3058
#>   0.6654  0.2916
#>   0.7498 -0.6295
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.7573  0.1731  0.0354  0.1422  0.3543  0.1588 -0.0892  0.6438 -0.1885
#>   0.1593 -0.3228 -0.1825 -0.3892  0.0022 -0.3674  0.5474 -0.5677 -0.4533
#>   0.5549  0.0248  0.7591 -0.1113 -0.0442  0.3874  0.0936  0.7376 -0.0319
#> 
#> Columns 10 to 18 -0.5499 -0.3711 -0.5549  0.4391  0.2682 -0.0212  0.3443 -0.7984 -0.5581
#>  -0.0562  0.5068 -0.4604 -0.1814 -0.2306 -0.0156  0.5692 -0.1581  0.1102
#>  -0.5079  0.4050  0.0724 -0.4349  0.3182  0.5683  0.4202 -0.3914 -0.5061
#> 
#> Columns 19 to 20  0.1426  0.6899
#>   0.3148 -0.6553
#>   0.5437 -0.1126
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2019  0.0507 -0.2220  0.1279  0.2241  0.2069 -0.3434  0.3685 -0.0494
#>   0.1938  0.0974  0.4094 -0.0524  0.2909  0.0967 -0.0844  0.3267  0.2415
#>   0.4412 -0.2277  0.4586  0.0129  0.2058  0.7124 -0.3016  0.6423  0.3188
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5686 -0.3729  0.4806  0.4354 -0.3178 -0.4502  0.3135 -0.2655 -0.3827
#>  -0.3230 -0.0315 -0.5130  0.7269 -0.0712  0.3497 -0.3745 -0.2678  0.3377
#>  -0.1251 -0.5145  0.2236  0.2684 -0.4593 -0.7627  0.4362 -0.2025 -0.3244
#> 
#> Columns 10 to 18 -0.1504  0.1657 -0.3040  0.0803 -0.0342 -0.0720  0.2219  0.4238  0.3304
#>   0.1196 -0.0174  0.5439  0.0363 -0.4162 -0.3879 -0.4704  0.0304 -0.2291
#>   0.3939  0.3248  0.2368  0.2251  0.3169  0.1779 -0.0677  0.6910 -0.2257
#> 
#> Columns 19 to 20  0.7220 -0.6030
#>  -0.1872  0.1560
#>   0.3687 -0.3104
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4331  0.2762  0.4080  0.2474  0.2817  0.3943 -0.0279 -0.0837  0.1898
#>   0.2797  0.2358  0.4579  0.2167 -0.0794  0.3587  0.1380  0.7581 -0.3067
#>   0.4471  0.3720  0.4718  0.3515  0.2126 -0.1941  0.5152 -0.2621 -0.2101
#> 
#> Columns 10 to 18 -0.5500 -0.1865  0.0200 -0.4140 -0.0152  0.0524  0.1650 -0.2044 -0.0009
#>  -0.4206 -0.3855 -0.2940 -0.1202  0.3373  0.0906  0.4173 -0.1779 -0.3149
#>  -0.3000 -0.0017 -0.1776  0.0166 -0.3719  0.1807  0.1471 -0.1685 -0.0685
#> 
#> Columns 19 to 20  0.7458  0.3621
#>   0.5129  0.1210
#>   0.7506 -0.1684
#> [ CPUFloatType{2,3,20} ]
#>