RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4695 -0.8002  0.1885 -0.8751 -0.7034  0.2117 -0.3618  0.9756 -0.3344
#>  -0.8706 -0.9849  0.6463  0.6464  0.6348  0.5558 -0.6260 -0.2442  0.0764
#>   0.8837 -0.6507 -0.0458 -0.4230 -0.8226  0.3012  0.4851  0.2766 -0.6887
#> 
#> Columns 10 to 18 -0.4947 -0.7046  0.3976  0.2252  0.2420 -0.4362 -0.3515 -0.3997 -0.8397
#>  -0.7801  0.4259  0.8495 -0.1231  0.3644  0.2500 -0.5995 -0.9658 -0.1638
#>  -0.2865 -0.6530  0.6823  0.7068 -0.4452 -0.4483  0.7054 -0.0629 -0.1632
#> 
#> Columns 19 to 20  0.7245 -0.8964
#>  -0.1125  0.7018
#>  -0.0542 -0.0220
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0223 -0.5486  0.5605 -0.3070  0.3308  0.6989  0.1066 -0.1177 -0.2247
#>   0.0744  0.3822  0.7441  0.1541 -0.0002  0.4940 -0.2828 -0.1131  0.3311
#>  -0.0729 -0.2731  0.4433 -0.6606  0.4918  0.0853  0.4106  0.1050 -0.2234
#> 
#> Columns 10 to 18 -0.7510 -0.0545 -0.2392  0.6274  0.5972 -0.4019  0.6629 -0.2714  0.7326
#>  -0.0848  0.5067  0.3247  0.3949 -0.0482 -0.5369  0.2698  0.0601  0.7486
#>  -0.6693 -0.0227  0.2061  0.3254  0.3003 -0.2484 -0.1033 -0.1679  0.4801
#> 
#> Columns 19 to 20  0.3560 -0.3540
#>  -0.0563 -0.5215
#>   0.2081 -0.2669
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2169  0.4132  0.5581 -0.1040 -0.0836  0.7201 -0.1237  0.1331  0.0352
#>  -0.2280  0.4150  0.2863 -0.6579 -0.0520  0.2198 -0.5595  0.6728 -0.1996
#>  -0.1998 -0.3936  0.1950 -0.2980 -0.1825  0.2407 -0.1868  0.2108 -0.0281
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0251 -0.5405 -0.1625 -0.2458  0.6160  0.0980  0.2380  0.1008 -0.3173
#>   0.0398 -0.2235 -0.1832  0.2980  0.4954 -0.0354 -0.2101  0.2154 -0.2550
#>   0.2440  0.7041  0.1085 -0.3616 -0.3660 -0.3295  0.6815  0.7359 -0.0668
#> 
#> Columns 10 to 18 -0.1778  0.0015  0.6743 -0.1874  0.4216  0.0009 -0.2120  0.2361 -0.6663
#>   0.4410  0.1818  0.6779 -0.4938 -0.0346  0.0817 -0.2568 -0.2889 -0.2887
#>   0.7794  0.4581 -0.9207 -0.0087 -0.1867  0.4034 -0.8272 -0.4889 -0.1082
#> 
#> Columns 19 to 20  0.2927 -0.8326
#>   0.5042 -0.8266
#>  -0.7694  0.6736
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2812  0.0530  0.3673  0.4481 -0.2961  0.8113  0.1314  0.0906 -0.1959
#>   0.1195 -0.0172  0.4905 -0.1150 -0.2780  0.6238 -0.2586  0.3174 -0.1729
#>  -0.5784  0.1126  0.1010 -0.3130 -0.6120 -0.0290  0.4507 -0.1915 -0.3148
#> 
#> Columns 10 to 18  0.1334 -0.0636 -0.1858  0.5320 -0.0528 -0.0317  0.2752  0.1403  0.3068
#>  -0.1172 -0.2938  0.0589  0.6755 -0.0073 -0.2155  0.2455  0.0391  0.0664
#>  -0.3536 -0.2411  0.4145  0.3922 -0.2659 -0.4459  0.1730  0.1224  0.2717
#> 
#> Columns 19 to 20 -0.4052  0.0193
#>  -0.3044 -0.3384
#>   0.1615 -0.5219
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>