RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2779 -0.2367  0.3003  0.6365  0.1520 -0.0751 -0.0145 -0.5058 -0.5387
#>  -0.0097  0.9426  0.7419  0.8389 -0.2636 -0.7972 -0.1346  0.7420  0.3002
#>   0.5742  0.8544 -0.3964  0.4319 -0.7011 -0.4899  0.6257 -0.1809 -0.1676
#> 
#> Columns 10 to 18 -0.8145  0.2086 -0.2647  0.1858 -0.2513 -0.4628  0.3745  0.5044  0.2990
#>  -0.1551  0.2347  0.9207  0.2561  0.1051 -0.7118 -0.2153 -0.1846 -0.4765
#>  -0.3960  0.0396  0.7293 -0.5292  0.6998  0.0129  0.4947 -0.2041  0.4127
#> 
#> Columns 19 to 20 -0.0981  0.4791
#>   0.1984  0.3323
#>   0.4136  0.6017
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1471  0.3251 -0.1074  0.1193  0.3082  0.3618 -0.5254  0.3907  0.5207
#>   0.0598  0.5572 -0.0128 -0.0008  0.1722 -0.7051 -0.2504  0.0899 -0.4612
#>  -0.0417 -0.0150 -0.5335  0.4929  0.2797  0.0517 -0.1503  0.6661  0.6052
#> 
#> Columns 10 to 18 -0.1611 -0.2154  0.5796  0.3820  0.4906 -0.4369 -0.5035  0.3098  0.1667
#>   0.5803  0.1102  0.2435 -0.1607 -0.1370 -0.3798 -0.6209  0.7193  0.4366
#>   0.0825 -0.6752 -0.0319 -0.8204 -0.0534 -0.4932 -0.1140  0.6582  0.8693
#> 
#> Columns 19 to 20  0.3992  0.0651
#>   0.3165 -0.7877
#>   0.1254  0.3578
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3727  0.3595  0.0408  0.2531  0.1000 -0.2891 -0.0933 -0.0980  0.0490
#>  -0.0862  0.4402  0.1364  0.3861  0.1314 -0.5130 -0.2922 -0.3160 -0.3076
#>   0.1490  0.1579  0.4299  0.5505  0.0158 -0.4327  0.1342 -0.7029  0.1694
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7413  0.7153  0.2072  0.1090 -0.6731  0.5400 -0.6902 -0.1326  0.3897
#>   0.2322 -0.2455 -0.3563 -0.3869  0.5968 -0.1141  0.0528  0.1416 -0.2795
#>  -0.3769 -0.0267  0.3631  0.0573  0.3728 -0.4676  0.2154 -0.7030  0.4604
#> 
#> Columns 10 to 18  0.2585 -0.6178 -0.5502  0.1975  0.4501  0.0747 -0.2505 -0.5331  0.3748
#>   0.2435  0.6281 -0.5390  0.3220  0.1267 -0.3005  0.3559 -0.1659  0.2539
#>   0.0250  0.5813 -0.7373  0.3833  0.4370 -0.2557  0.8465 -0.0999 -0.3850
#> 
#> Columns 19 to 20 -0.4101 -0.1528
#>  -0.0034 -0.1483
#>   0.6805  0.1271
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0848  0.3229  0.1394  0.1184 -0.0638 -0.1372 -0.1300  0.0823  0.0353
#>  -0.0244  0.5063 -0.0045  0.1693  0.0577 -0.4599 -0.1776 -0.4414 -0.2362
#>   0.2310  0.4256  0.2241 -0.0179  0.1687 -0.5674  0.1431 -0.2836 -0.5558
#> 
#> Columns 10 to 18 -0.2317 -0.0773 -0.0766 -0.3882 -0.2598  0.2731  0.0841  0.0591 -0.3892
#>  -0.4185 -0.0984  0.1783 -0.2218 -0.2251 -0.1807 -0.0352  0.6498  0.5597
#>   0.4625  0.0568  0.1115 -0.3778 -0.4372 -0.0359 -0.3462  0.6957  0.2933
#> 
#> Columns 19 to 20 -0.0959 -0.3789
#>   0.2346 -0.1634
#>   0.0530 -0.6083
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>