RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2546  0.0221 -0.1485 -0.0250 -0.3161 -0.3763  0.6158 -0.8434 -0.8544
#>  -0.7976  0.2706 -0.6370  0.1773 -0.7202  0.9476 -0.5365  0.2477  0.2985
#>  -0.7330  0.2765  0.0752  0.7916 -0.6562  0.1643 -0.4983  0.1979  0.2319
#> 
#> Columns 10 to 18 -0.4154 -0.8657  0.0582  0.1426  0.1576 -0.6216  0.7955  0.7741  0.7023
#>  -0.5211  0.4193  0.5174 -0.2688 -0.7228 -0.0420  0.0990  0.3385  0.7532
#>  -0.6787  0.4560 -0.2490  0.2606 -0.5927  0.2153 -0.4186  0.1694 -0.3440
#> 
#> Columns 19 to 20  0.5161 -0.8876
#>  -0.2551 -0.0698
#>   0.4836  0.6738
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4746  0.5676 -0.1598  0.3521  0.0357  0.5592  0.7082  0.1584  0.0971
#>  -0.3630  0.7125 -0.0059 -0.6139 -0.3430 -0.2859  0.7355  0.3796  0.3808
#>  -0.7304  0.6825 -0.1184 -0.3779 -0.6483  0.0575  0.2347  0.4466  0.3957
#> 
#> Columns 10 to 18  0.0429  0.1367  0.3868 -0.5010 -0.1716  0.4521  0.2109  0.3496  0.2003
#>  -0.0171 -0.3551 -0.2263  0.0636  0.2385 -0.4423 -0.2559  0.1413  0.0341
#>  -0.0060  0.1274 -0.4591  0.3924  0.1003 -0.2803 -0.4654  0.2185 -0.3418
#> 
#> Columns 19 to 20  0.5992  0.0620
#>   0.4193 -0.0196
#>   0.1145  0.5248
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.7405  0.2544  0.2205 -0.4981 -0.4481  0.3407  0.1865 -0.1521  0.2751
#>  -0.1793  0.6760 -0.6325 -0.1795 -0.3994 -0.1543 -0.4053  0.0262 -0.3627
#>  -0.3084  0.4062 -0.2252 -0.1469 -0.3720 -0.0185 -0.0628  0.2901 -0.2321
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2485 -0.2926  0.1090  0.1382 -0.5231  0.2826 -0.3996  0.8478 -0.8451
#>  -0.2325 -0.1023  0.2241  0.6820 -0.0119  0.8195 -0.6730 -0.4823 -0.8230
#>  -0.7039  0.0048  0.0053 -0.1656  0.2964  0.3017  0.0709  0.1344 -0.1703
#> 
#> Columns 10 to 18 -0.6042  0.5276 -0.3986 -0.0638  0.0361  0.5392  0.5168  0.4671  0.1946
#>  -0.7552 -0.1886 -0.4707  0.0188  0.0634  0.3792  0.4678  0.3392  0.2784
#>   0.1257  0.3305  0.1139  0.2093 -0.2697  0.5119  0.4285 -0.1495  0.5091
#> 
#> Columns 19 to 20  0.8167  0.3185
#>  -0.1471  0.0926
#>  -0.2914 -0.2976
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5702 -0.0265  0.1087 -0.4187 -0.0599  0.5310  0.3814 -0.0484 -0.0427
#>  -0.2716 -0.0830  0.0737 -0.3383  0.1435  0.0029  0.2599  0.1076 -0.1526
#>  -0.6205  0.7195 -0.1830 -0.5150 -0.2677 -0.1832 -0.1918  0.3992  0.1915
#> 
#> Columns 10 to 18 -0.1446  0.2899  0.3653  0.2761  0.0444  0.2264 -0.2531  0.4629  0.4708
#>  -0.3426 -0.0487  0.1589  0.2179 -0.2158  0.1291 -0.1901  0.1632  0.4293
#>   0.0316 -0.2615 -0.2996 -0.0626 -0.0797 -0.1907  0.1953 -0.0196 -0.2929
#> 
#> Columns 19 to 20 -0.0808  0.2984
#>   0.2124  0.0666
#>   0.0907 -0.1149
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>