RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6672 -0.7015 -0.2506  0.7103 -0.1327 -0.0916 -0.5275  0.0242 -0.8767
#>   0.1328 -0.5158  0.4956  0.5229 -0.0046  0.3718 -0.0414 -0.5988  0.8271
#>  -0.6509  0.8422 -0.5292 -0.8126 -0.5383  0.5916 -0.4797 -0.4408  0.2166
#> 
#> Columns 10 to 18 -0.6946 -0.2242 -0.7150  0.1174  0.2975 -0.7947 -0.0460  0.3469  0.1848
#>  -0.3165  0.5779 -0.8185  0.3524 -0.8253  0.2850  0.0712 -0.4209  0.3970
#>   0.0651 -0.1419  0.4455  0.9334  0.0343 -0.0133  0.0695 -0.5590  0.2187
#> 
#> Columns 19 to 20 -0.3740 -0.1536
#>  -0.6674  0.3819
#>   0.5144  0.6592
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1915 -0.2124 -0.1838  0.6231 -0.0847 -0.1057 -0.5064  0.0575 -0.4860
#>  -0.5659 -0.4591 -0.4386  0.5499 -0.1975 -0.1206 -0.3582 -0.4750 -0.2716
#>  -0.4744 -0.5692  0.5839  0.5106 -0.6622  0.2589  0.4601  0.0987  0.3195
#> 
#> Columns 10 to 18 -0.1403 -0.1389 -0.2695 -0.2993 -0.5814 -0.0406 -0.4193  0.2924 -0.1358
#>  -0.1746  0.2756  0.0481  0.0691  0.1485 -0.5900 -0.6413 -0.1038  0.0694
#>  -0.5885 -0.4393  0.1935 -0.4560  0.2384 -0.4360 -0.0839 -0.7321  0.6849
#> 
#> Columns 19 to 20  0.0619  0.0090
#>   0.0933  0.2145
#>  -0.2715 -0.3123
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2286 -0.3541  0.4000  0.6791 -0.2418 -0.4153 -0.3109 -0.0229 -0.3426
#>  -0.5395 -0.4477  0.3236  0.4536 -0.0068 -0.1421 -0.2904  0.0847 -0.3430
#>   0.1946 -0.0860 -0.2615 -0.1307 -0.4463 -0.0551 -0.2704 -0.5765 -0.2200
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0422  0.7061 -0.4162  0.7530 -0.7170 -0.3496 -0.1357  0.2456  0.1516
#>   0.5842  0.0439  0.3068  0.2843 -0.5837  0.5360  0.4571  0.1491  0.2873
#>   0.2310  0.5974 -0.2138 -0.1865  0.3222  0.1776  0.3378  0.4641 -0.6285
#> 
#> Columns 10 to 18  0.6199  0.0682  0.7333 -0.0938  0.6835  0.0724  0.6901 -0.5221  0.2344
#>   0.3155  0.4003  0.4242 -0.1723 -0.0608  0.2099 -0.1028  0.0628  0.0430
#>  -0.4278 -0.5906 -0.3467  0.5439 -0.2721  0.2647  0.7588 -0.0917  0.7146
#> 
#> Columns 19 to 20  0.5372 -0.3415
#>   0.3786  0.8045
#>   0.6219 -0.0542
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3163 -0.4464  0.1893  0.5896 -0.0693 -0.2983 -0.2544 -0.0262 -0.0852
#>  -0.3688 -0.4501  0.5020  0.4293 -0.6388  0.2226 -0.0161 -0.2035 -0.0774
#>   0.0478 -0.2841 -0.1517  0.1845 -0.0751 -0.1456 -0.2640 -0.2799 -0.0483
#> 
#> Columns 10 to 18 -0.1422 -0.0772 -0.1400 -0.3988  0.2534 -0.1843 -0.0500  0.2171 -0.1352
#>  -0.4267 -0.0740  0.0714 -0.2736 -0.2548 -0.2421 -0.0533 -0.2543  0.5794
#>  -0.6043  0.0597  0.2919  0.1353  0.2277 -0.4265 -0.5458 -0.0255 -0.3518
#> 
#> Columns 19 to 20 -0.2933  0.2812
#>  -0.2906 -0.0887
#>   0.3287  0.2020
#> [ CPUFloatType{2,3,20} ]
#>