RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0097 -0.5045  0.2320  0.5998  0.6707 -0.3347  0.8717 -0.2740  0.1408
#>   0.1807  0.6981  0.4005 -0.4251 -0.7445  0.2844  0.7787 -0.0044  0.1380
#>   0.3364 -0.8042 -0.5250 -0.2637 -0.4474 -0.3187  0.5633 -0.0153 -0.2968
#> 
#> Columns 10 to 18  0.4656  0.4261 -0.0502  0.7443  0.7051  0.2039 -0.7314  0.3883  0.4622
#>   0.3448 -0.2336  0.1185  0.2218  0.3597  0.2027  0.3527 -0.2446  0.1895
#>  -0.1187 -0.1952  0.7009  0.8571 -0.2260  0.1764 -0.6245  0.0821 -0.0447
#> 
#> Columns 19 to 20  0.4672 -0.8740
#>  -0.2559 -0.1972
#>  -0.3741 -0.2312
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0502 -0.1979 -0.2404  0.2097 -0.5076  0.2622  0.1454 -0.1831  0.2128
#>   0.2300 -0.0468 -0.0720  0.2459  0.3618 -0.3175 -0.0119 -0.4021  0.2086
#>  -0.1442 -0.3964 -0.5110  0.2645  0.0806  0.3425  0.6001 -0.3809 -0.2641
#> 
#> Columns 10 to 18  0.1463  0.1864 -0.0739  0.1154 -0.0092 -0.0831  0.2963  0.6689  0.1369
#>  -0.5127  0.5651 -0.5001 -0.0322  0.1936 -0.1201 -0.3186  0.3806  0.3900
#>  -0.0315  0.2194  0.1666  0.2557 -0.1781 -0.1340 -0.1295  0.3235  0.2907
#> 
#> Columns 19 to 20  0.0742  0.3064
#>  -0.4028  0.3303
#>  -0.4343  0.3049
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1740 -0.5236 -0.5186  0.1390 -0.0171  0.0583  0.6904  0.2630 -0.0409
#>   0.2813 -0.0542 -0.0043 -0.0625 -0.1352  0.1359  0.3874 -0.0481  0.0800
#>  -0.0334 -0.1896 -0.3981  0.3126 -0.3327  0.3170  0.5067  0.0522 -0.3226
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4524  0.4506  0.1316 -0.3215 -0.0199 -0.4286 -0.6492  0.1285 -0.2514
#>  -0.0872 -0.3364  0.4178  0.4796 -0.5253 -0.3154 -0.5564 -0.3937  0.4331
#>   0.7226  0.3420 -0.0767 -0.0522  0.2922 -0.7643 -0.4221  0.2106 -0.5154
#> 
#> Columns 10 to 18 -0.2365  0.3380 -0.3327  0.1692  0.4935  0.2827 -0.4936  0.5340  0.1586
#>  -0.3233 -0.1393  0.4022 -0.7970  0.5930  0.3624  0.0819 -0.2645 -0.3899
#>  -0.7149 -0.0048  0.4174 -0.4868 -0.2244  0.5786 -0.2956  0.3367 -0.1774
#> 
#> Columns 19 to 20 -0.2383 -0.3674
#>   0.1007 -0.5005
#>  -0.2449  0.0867
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0394 -0.2836 -0.4533  0.2454 -0.1685  0.3381  0.5301 -0.0589 -0.0254
#>   0.1837  0.4322  0.1873  0.0922 -0.1565  0.3345  0.6365 -0.0394 -0.0320
#>   0.1404 -0.0863 -0.5576  0.2769 -0.0280  0.0594  0.4602  0.1701  0.0814
#> 
#> Columns 10 to 18 -0.0371  0.3840  0.4584  0.2457  0.0363 -0.0483 -0.3419  0.4960  0.1402
#>   0.2126 -0.3499 -0.0231  0.3037  0.4736  0.1924 -0.3468  0.0946  0.0309
#>  -0.1126  0.1921 -0.0763  0.2587 -0.0518 -0.4425 -0.2708  0.4705  0.1242
#> 
#> Columns 19 to 20  0.2186  0.1988
#>  -0.0059 -0.4933
#>  -0.3250 -0.3887
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>