RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9587 -0.3390  0.6635 -0.4425 -0.2474  0.1528  0.3009  0.0508 -0.3612
#>  -0.4943  0.4810  0.0429 -0.2638  0.5849  0.0534 -0.4797 -0.4887  0.3902
#>  -0.5339  0.5359  0.4051  0.5988 -0.8264 -0.0587  0.4100 -0.3253  0.0632
#> 
#> Columns 10 to 18 -0.6679 -0.9124 -0.1680 -0.1338  0.8372 -0.5988  0.5935  0.0945  0.3280
#>   0.2638  0.6329  0.6330  0.0168  0.9107  0.8654  0.1993 -0.4657  0.6764
#>   0.5247 -0.9630 -0.4832 -0.5224 -0.6381 -0.3744 -0.5333  0.3720  0.9120
#> 
#> Columns 19 to 20 -0.5269 -0.0035
#>  -0.5007  0.1151
#>   0.2156  0.6051
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4656  0.0710  0.3042  0.2470 -0.4482  0.2285 -0.3846  0.1623  0.1840
#>   0.0307  0.2334  0.2829 -0.0718 -0.5991 -0.1443  0.0412 -0.2226 -0.5143
#>  -0.5561  0.2717  0.4007  0.3848 -0.6147  0.1484 -0.8215  0.0505  0.3386
#> 
#> Columns 10 to 18  0.7959  0.1354  0.0825 -0.1349 -0.1345 -0.3014 -0.3833 -0.0076  0.6166
#>   0.7799 -0.4046 -0.1440 -0.3140  0.3378  0.3078 -0.0728  0.2421 -0.3706
#>   0.0115 -0.3659 -0.5727 -0.0870 -0.2615 -0.5473 -0.1269  0.0148  0.1618
#> 
#> Columns 19 to 20  0.2909 -0.0470
#>  -0.4291  0.5814
#>  -0.0628 -0.3045
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3318 -0.3650  0.2518  0.3870 -0.2431 -0.4290 -0.2495  0.1812 -0.3419
#>   0.2367  0.1450  0.2728  0.4157  0.0342 -0.4162 -0.3936  0.3876 -0.1323
#>   0.5500  0.1408  0.3401 -0.2316 -0.2092 -0.3768 -0.5803  0.2658 -0.1934
#> 
#> Columns 10 to 18  0.4013 -0.1620 -0.5450 -0.4930  0.0526  0.2163 -0.7187 -0.1574  0.5179
#>   0.7084  0.3030 -0.1365 -0.5498  0.0194  0.5889  0.0333  0.0474  0.7875
#>   0.7616  0.3934 -0.0460 -0.3610  0.1677  0.2190 -0.4397 -0.2886  0.5592
#> 
#> Columns 19 to 20 -0.0072  0.4405
#>   0.0481 -0.1293
#>   0.1578  0.2117
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.1393  0.2621  0.4567  0.0909 -0.6019  0.2075 -0.6397 -0.2491 -0.2129
#>   0.0641  0.2894  0.2196  0.4086 -0.7564 -0.0792 -0.3913 -0.0799 -0.4999
#>   0.1594 -0.1796  0.1001  0.4511 -0.5139  0.0044 -0.1935  0.0657 -0.7292
#> 
#> Columns 10 to 18  0.7355  0.0603 -0.0467 -0.4530  0.0835  0.0279 -0.2120  0.3682  0.2352
#>   0.7716 -0.4408 -0.2666  0.0184  0.1913  0.1697 -0.3829  0.5277  0.3508
#>   0.7186  0.2023 -0.4439 -0.4143  0.3262  0.1947 -0.6055  0.3454  0.2707
#> 
#> Columns 19 to 20 -0.2769 -0.1638
#>  -0.1852 -0.0246
#>   0.0896  0.2570
#> 
#> (5,.,.) = 
#>  Columns 1 to 9 -0.2939 -0.0345  0.2820  0.4670 -0.7628 -0.2658 -0.3737 -0.0755 -0.2024
#>  -0.1936  0.0672  0.1773  0.2520 -0.2834 -0.3211 -0.5540 -0.0945 -0.2840
#>   0.1721  0.1677  0.4495  0.2505 -0.4911  0.0897 -0.4675 -0.3366 -0.2284
#> 
#> Columns 10 to 18  0.5860 -0.3120 -0.4980 -0.1171  0.0873 -0.0560 -0.2166  0.5246  0.6399
#>   0.6032  0.1032 -0.3736 -0.3454  0.2099  0.5395 -0.0583  0.2401  0.5146
#>   0.8791  0.2460 -0.0361 -0.3574  0.2179  0.4373 -0.4803  0.7651  0.4143
#> 
#> Columns 19 to 20  0.2914 -0.3160
#>   0.0425  0.2589
#>   0.4216  0.0034
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.3397  0.1766 -0.3330  0.3067  0.1849 -0.5877 -0.8002  0.6076  0.1962
#>   0.0935 -0.5411 -0.4536 -0.4104 -0.4696  0.7975 -0.3846 -0.1947 -0.1160
#>   0.4956 -0.5389 -0.3684  0.3573 -0.1905 -0.2440  0.6433 -0.4237 -0.4549
#> 
#> Columns 10 to 18 -0.7556 -0.1609 -0.0061  0.6834  0.0599 -0.6043  0.3497 -0.0773  0.3177
#>  -0.4388 -0.0717 -0.1671 -0.5604  0.6176  0.6608  0.3035 -0.2334 -0.4388
#>  -0.1985 -0.3392  0.2658 -0.5242 -0.4235  0.0932  0.0189  0.2252 -0.0019
#> 
#> Columns 19 to 20  0.3045  0.0872
#>   0.2692 -0.1228
#>   0.8302  0.2420
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2939 -0.0345  0.2820  0.4670 -0.7628 -0.2658 -0.3737 -0.0755 -0.2024
#>  -0.1936  0.0672  0.1773  0.2520 -0.2834 -0.3211 -0.5540 -0.0945 -0.2840
#>   0.1721  0.1677  0.4495  0.2505 -0.4911  0.0897 -0.4675 -0.3366 -0.2284
#> 
#> Columns 10 to 18  0.5860 -0.3120 -0.4980 -0.1171  0.0873 -0.0560 -0.2166  0.5246  0.6399
#>   0.6032  0.1032 -0.3736 -0.3454  0.2099  0.5395 -0.0583  0.2401  0.5146
#>   0.8791  0.2460 -0.0361 -0.3574  0.2179  0.4373 -0.4803  0.7651  0.4143
#> 
#> Columns 19 to 20  0.2914 -0.3160
#>   0.0425  0.2589
#>   0.4216  0.0034
#> [ CPUFloatType{2,3,20} ]
#>