RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1054 -0.5115  0.2681 -0.4143  0.6832 -0.9526 -0.8451 -0.6179  0.5689
#>  -0.6397  0.4763 -0.1843 -0.5796 -0.3844  0.1461 -0.2647  0.6388 -0.7167
#>   0.7082  0.3591  0.6999 -0.0805 -0.7943  0.1960  0.4116  0.6571 -0.0142
#> 
#> Columns 10 to 18 -0.7758 -0.5427  0.5450  0.7241 -0.6650 -0.0782  0.4902  0.3657  0.6062
#>   0.8100 -0.5767  0.2122  0.3805  0.7574  0.6055 -0.7987  0.7254 -0.5656
#>   0.0138 -0.2356 -0.1305 -0.0705 -0.2719  0.4994 -0.8123  0.2795 -0.2818
#> 
#> Columns 19 to 20  0.4061  0.0695
#>  -0.7710  0.7705
#>   0.5710  0.6093
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0804  0.3174 -0.6888  0.5332  0.1847  0.0227  0.1545 -0.2843  0.3060
#>   0.7107 -0.6424  0.3943 -0.7251 -0.2163 -0.0488 -0.1424 -0.4056 -0.0460
#>   0.7034 -0.1454 -0.2360  0.4767 -0.0859 -0.4702 -0.4411 -0.4143  0.3556
#> 
#> Columns 10 to 18  0.4232  0.0583 -0.1184  0.3786  0.2564  0.0168 -0.3011  0.2926  0.6812
#>  -0.1795 -0.6148  0.4587  0.3564 -0.6472 -0.1959 -0.3892 -0.1961 -0.6444
#>  -0.1440 -0.4948 -0.0571 -0.0610 -0.0938  0.1149  0.1807 -0.6778  0.4499
#> 
#> Columns 19 to 20 -0.5760  0.0870
#>   0.4772  0.3093
#>   0.0184 -0.3414
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0809  0.3861 -0.0582 -0.0365 -0.1314 -0.4983 -0.2984  0.3763 -0.2341
#>  -0.0351 -0.2824 -0.6650  0.4051 -0.0740  0.1229 -0.3845 -0.4704  0.0837
#>  -0.1299  0.3260 -0.6483  0.2289 -0.2582  0.0732 -0.5545  0.5417  0.0988
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2272  0.7976 -0.4344  0.1722 -0.5105  0.4327 -0.7351  0.6828  0.0039
#>   0.8735  0.6511  0.4238  0.3356  0.6102  0.2351  0.0551  0.5092 -0.2888
#>  -0.5986  0.3657 -0.3868 -0.2831 -0.2994  0.7899 -0.2987  0.6600  0.1141
#> 
#> Columns 10 to 18  0.5333  0.4318  0.4892 -0.3862 -0.2377  0.4710  0.3834 -0.2699 -0.3060
#>  -0.3715 -0.6862 -0.0997 -0.3900  0.0841 -0.2477  0.7893  0.0686  0.0620
#>   0.6250  0.5278  0.1592  0.2764 -0.1872  0.3416 -0.1996  0.3699  0.3647
#> 
#> Columns 19 to 20 -0.2015  0.2797
#>   0.0432 -0.1246
#>  -0.3521  0.0181
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1910  0.3369 -0.0652  0.1103 -0.6070 -0.4070  0.0507  0.1167  0.4215
#>   0.3132  0.0356 -0.0715 -0.1634 -0.2544  0.1688 -0.2520  0.5957 -0.3035
#>   0.1429 -0.0682 -0.4441 -0.0040 -0.6382 -0.1742 -0.1075  0.0032  0.4300
#> 
#> Columns 10 to 18  0.2362 -0.3570  0.3054 -0.3460 -0.0094 -0.3505  0.0238  0.0946  0.1848
#>   0.4297 -0.2660  0.3099 -0.0154  0.2137  0.0648 -0.4015 -0.2670 -0.3589
#>  -0.1792 -0.1380  0.0164  0.0392 -0.2655 -0.0959 -0.4215 -0.1110 -0.4445
#> 
#> Columns 19 to 20 -0.1002  0.1634
#>  -0.0098 -0.0004
#>   0.0382  0.4542
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>