RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1084  0.3774  0.2897  0.7969 -0.4613 -0.1375  0.4145  0.6782  0.8935
#>  -0.6685  0.3390  0.6508  0.6319  0.8159 -0.2216  0.5974  0.9386  0.2852
#>  -0.0225  0.4994  0.2839 -0.1598  0.6401 -0.5176  0.5185 -0.2931 -0.0013
#> 
#> Columns 10 to 18 -0.7742  0.7912  0.3762  0.2345  0.4036  0.7390  0.4369  0.7904 -0.0972
#>  -0.2673 -0.4963 -0.1010  0.5082 -0.1845 -0.0572  0.6143  0.2983 -0.7093
#>   0.3466  0.2641 -0.5976 -0.4964 -0.1208 -0.5897  0.2612 -0.4000 -0.2731
#> 
#> Columns 19 to 20  0.8831 -0.3418
#>  -0.0069 -0.9038
#>   0.3932 -0.5623
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0529  0.6240  0.6361  0.7833  0.0747  0.3809  0.1082  0.5868  0.0407
#>  -0.1094  0.4152  0.6018  0.5864  0.0003  0.1487  0.5594  0.4503 -0.3323
#>   0.6156  0.0562  0.4749  0.0159 -0.3741  0.0161  0.6816  0.2124 -0.1097
#> 
#> Columns 10 to 18  0.0301  0.4105  0.0459 -0.2608  0.6151  0.5597  0.1994 -0.0547  0.7041
#>  -0.3547 -0.0153 -0.1686 -0.1034  0.4315  0.3041  0.4498 -0.1331  0.1215
#>  -0.5556  0.2118  0.1159 -0.2205  0.3333  0.1859  0.0598 -0.1802 -0.2301
#> 
#> Columns 19 to 20  0.7274  0.2542
#>   0.0588 -0.1187
#>   0.3083 -0.0193
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1864  0.5881  0.2793  0.1701  0.4799  0.2421 -0.1932  0.2962 -0.1952
#>   0.1153  0.2980  0.0277  0.5364  0.1198  0.4293  0.0780  0.4103 -0.1326
#>   0.0809 -0.0472  0.4777  0.2492  0.0308  0.7123 -0.1634  0.7061 -0.1043
#> 
#> Columns 10 to 18 -0.4366 -0.2839 -0.4780 -0.3088 -0.2057  0.0005  0.2435 -0.5711  0.1500
#>  -0.2949 -0.4826 -0.2748 -0.2543  0.2045 -0.1584  0.0024 -0.2949 -0.0116
#>  -0.5491 -0.4128 -0.3481 -0.3073  0.2899 -0.0845  0.2121 -0.2246 -0.2509
#> 
#> Columns 19 to 20  0.8216 -0.0011
#>   0.4984  0.2203
#>   0.5619 -0.0690
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.3061  0.0958  0.3000 -0.0347 -0.1344 -0.0768  0.2000  0.0571 -0.0502
#>   0.3380  0.4580  0.1199  0.4718  0.1823  0.2526  0.4643  0.0962 -0.2931
#>   0.6304  0.2107  0.1856  0.1326 -0.1200  0.4645  0.1989  0.3598  0.0426
#> 
#> Columns 10 to 18 -0.1924  0.0697 -0.1953 -0.3354 -0.1467 -0.0865  0.0752 -0.4369  0.0524
#>  -0.4680 -0.1950 -0.3005 -0.2166 -0.1167 -0.1073  0.0373 -0.0519  0.0808
#>  -0.2040 -0.0385 -0.3404 -0.2266 -0.0289 -0.2061 -0.2896 -0.5202 -0.1562
#> 
#> Columns 19 to 20  0.7140  0.0633
#>   0.7751  0.2233
#>   0.3564  0.1723
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.6578  0.0232  0.5084  0.3437 -0.0102  0.2330  0.5131  0.0541 -0.0927
#>   0.4510  0.0913  0.3092  0.5337  0.1974  0.1114  0.5691 -0.1046 -0.5297
#>   0.6200  0.3281  0.3775  0.2576  0.1257  0.3188  0.1789  0.6142 -0.1083
#> 
#> Columns 10 to 18 -0.1524  0.0730 -0.3861 -0.1675 -0.0150  0.0251 -0.1046 -0.3870  0.0553
#>  -0.2672 -0.2624 -0.2319  0.0015  0.1198  0.3474  0.1082 -0.2297  0.0309
#>  -0.4336 -0.2089 -0.5096  0.2524  0.2779 -0.1315  0.1764 -0.5567 -0.2169
#> 
#> Columns 19 to 20  0.5348  0.4251
#>   0.6230 -0.1151
#>   0.5614  0.3486
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.1306  0.0630 -0.0515 -0.2924 -0.0999 -0.7401  0.6148 -0.1446  0.0810
#>  -0.5493  0.1240  0.0775  0.2439  0.0221 -0.3468  0.3645 -0.4158  0.6457
#>   0.4518 -0.3932 -0.3992  0.8918  0.6219 -0.4087  0.2826 -0.2738  0.4347
#> 
#> Columns 10 to 18  0.5757 -0.1065 -0.6391  0.2611 -0.3880 -0.0748 -0.1064  0.1305 -0.5569
#>   0.7209  0.4715  0.0844  0.0223  0.1424  0.4875 -0.1123  0.3386 -0.4078
#>  -0.1123 -0.3192  0.1003 -0.4439 -0.3718 -0.4325 -0.3494  0.0328 -0.6454
#> 
#> Columns 19 to 20  0.3684 -0.0321
#>  -0.3940  0.4542
#>   0.2532 -0.0918
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6578  0.0232  0.5084  0.3437 -0.0102  0.2330  0.5131  0.0541 -0.0927
#>   0.4510  0.0913  0.3092  0.5337  0.1974  0.1114  0.5691 -0.1046 -0.5297
#>   0.6200  0.3281  0.3775  0.2576  0.1257  0.3188  0.1789  0.6142 -0.1083
#> 
#> Columns 10 to 18 -0.1524  0.0730 -0.3861 -0.1675 -0.0150  0.0251 -0.1046 -0.3870  0.0553
#>  -0.2672 -0.2624 -0.2319  0.0015  0.1198  0.3474  0.1082 -0.2297  0.0309
#>  -0.4336 -0.2089 -0.5096  0.2524  0.2779 -0.1315  0.1764 -0.5567 -0.2169
#> 
#> Columns 19 to 20  0.5348  0.4251
#>   0.6230 -0.1151
#>   0.5614  0.3486
#> [ CPUFloatType{2,3,20} ]
#>