RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3997  0.7120 -0.4973 -0.1341 -0.2115  0.8263  0.7727  0.9201  0.2067
#>   0.1940 -0.4119 -0.4361  0.0411 -0.2434  0.7737  0.5384  0.6834 -0.7905
#>  -0.0878  0.6705  0.0793 -0.2587 -0.1544 -0.6748  0.9656 -0.4491 -0.4869
#> 
#> Columns 10 to 18 -0.4061  0.6662 -0.3064  0.7642 -0.4539  0.3899 -0.6753  0.5038 -0.4567
#>   0.3253  0.7641 -0.7409  0.9234 -0.0844 -0.8945 -0.4543 -0.1554 -0.7042
#>  -0.2330 -0.6739 -0.4642 -0.1488  0.5713  0.7289  0.0751 -0.9527 -0.2316
#> 
#> Columns 19 to 20  0.4863  0.4375
#>   0.6129 -0.4574
#>   0.8227 -0.0545
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0785 -0.0176 -0.6793 -0.2732 -0.1760  0.2655  0.4079 -0.4646  0.3204
#>  -0.3772 -0.3944 -0.6011 -0.1246  0.3311  0.2639 -0.1426 -0.3771  0.4581
#>   0.3515 -0.1563 -0.5481  0.2763  0.0264  0.2259  0.3976  0.5796 -0.1104
#> 
#> Columns 10 to 18  0.0655  0.4764 -0.1292  0.1270 -0.6490 -0.3764 -0.0853 -0.5510 -0.4908
#>   0.1035  0.4903 -0.4093 -0.4381 -0.5479 -0.5206 -0.4886 -0.1088 -0.0018
#>   0.5497  0.5843 -0.1040  0.5300 -0.6275  0.2160 -0.3447 -0.2891 -0.7418
#> 
#> Columns 19 to 20 -0.1682 -0.6444
#>  -0.1356 -0.3917
#>   0.5344  0.3541
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1088 -0.3639 -0.3826  0.1475  0.5201  0.4829  0.0963 -0.0905  0.5963
#>  -0.2084 -0.3252 -0.1295 -0.2149 -0.0181  0.4088  0.0783 -0.3163  0.3495
#>   0.1379 -0.5681 -0.2114 -0.1895  0.1247  0.0678  0.2863 -0.3577  0.5279
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7497 -0.2515  0.3100  0.3013  0.5588 -0.1765 -0.9069  0.2829  0.6974
#>   0.3176  0.0801  0.7553  0.1418 -0.5550 -0.0279  0.3014  0.4918  0.3990
#>  -0.7622  0.4328 -0.6515 -0.0323  0.3885  0.3862  0.4530  0.6013  0.6516
#> 
#> Columns 10 to 18  0.1911  0.2980  0.1070  0.8630 -0.0699  0.2724 -0.7774 -0.1993  0.7437
#>  -0.5679  0.0032 -0.4750 -0.6113  0.3148  0.3084 -0.7486 -0.4061  0.6442
#>  -0.5852  0.4560  0.4431  0.6479 -0.3639  0.2441 -0.4993 -0.2683  0.6262
#> 
#> Columns 19 to 20 -0.4259 -0.3506
#>  -0.6386 -0.5488
#>  -0.4985  0.0977
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1929 -0.2092 -0.3109  0.1193  0.1048  0.4637  0.1665  0.0148  0.2619
#>  -0.3950  0.2599 -0.0903 -0.0545  0.0873 -0.1369  0.2323 -0.0869  0.6825
#>  -0.3035 -0.1196 -0.0718  0.0186 -0.3159  0.2680 -0.1749 -0.1777  0.2662
#> 
#> Columns 10 to 18 -0.0376  0.1854 -0.2050 -0.1225 -0.2271  0.4979 -0.2275 -0.3388  0.2525
#>   0.3966  0.2700  0.1882  0.4413 -0.6130  0.6599 -0.5136 -0.6152 -0.5183
#>  -0.2963  0.2863 -0.4218  0.0327 -0.4356  0.0570  0.0980 -0.4824  0.1349
#> 
#> Columns 19 to 20 -0.0245  0.1457
#>   0.3385  0.3926
#>  -0.4152 -0.4642
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>