RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2988  0.8523 -0.5005  0.6845 -0.0447  0.2209 -0.3012  0.8784 -0.5637
#>   0.7280  0.6728  0.0200  0.6394 -0.6731  0.4551  0.7604  0.9252 -0.4317
#>  -0.6168  0.1317 -0.5349  0.7928  0.5155 -0.3271  0.5303  0.3266  0.2454
#> 
#> Columns 10 to 18  0.7460 -0.8120 -0.3087  0.1661 -0.6322  0.1280  0.7219 -0.1256 -0.2297
#>   0.7943  0.7147 -0.4327 -0.0047  0.4036  0.5867  0.1198 -0.6136 -0.6285
#>  -0.3361  0.2795 -0.3587  0.5246 -0.8221  0.1591  0.0679 -0.2762  0.1201
#> 
#> Columns 19 to 20  0.3145 -0.2777
#>   0.4031 -0.4937
#>   0.6150 -0.6865
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1974  0.5454 -0.4540  0.1630 -0.0145 -0.2488  0.6765 -0.1259 -0.5820
#>   0.4125  0.6754 -0.0125  0.2724 -0.3248  0.2427 -0.2928  0.6258 -0.3456
#>  -0.2019  0.2021  0.6524  0.1044  0.2859 -0.5033  0.6863 -0.2220 -0.5490
#> 
#> Columns 10 to 18 -0.3699 -0.2689 -0.0837  0.1435 -0.0109 -0.1532  0.0191  0.0983  0.3119
#>  -0.7964 -0.4679 -0.2116  0.1702  0.5960 -0.1080 -0.2514 -0.2666  0.5889
#>  -0.3854 -0.1358 -0.2559 -0.6372  0.0527  0.2312  0.5646  0.3128 -0.1026
#> 
#> Columns 19 to 20  0.5516 -0.0843
#>   0.7681  0.3182
#>   0.2508 -0.4269
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2413  0.3122  0.2656  0.3366 -0.2187  0.2757  0.6146  0.3366 -0.0490
#>   0.1558 -0.1912  0.3595  0.4505  0.4339  0.3466  0.4544 -0.4700 -0.3860
#>   0.0220 -0.0331  0.4404  0.4039  0.0636  0.6270  0.2797  0.5700 -0.0880
#> 
#> Columns 10 to 18 -0.0792  0.1046  0.3088 -0.3501 -0.3219  0.1501  0.1721 -0.0493 -0.2556
#>   0.0108 -0.1614 -0.2790 -0.1680 -0.4998  0.1121 -0.5297 -0.2721  0.0150
#>  -0.6308 -0.5919 -0.1732 -0.5696  0.4135  0.1058  0.5133  0.0881 -0.5690
#> 
#> Columns 19 to 20  0.5849  0.1790
#>   0.6202 -0.4740
#>   0.3393 -0.0369
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.5922  0.3411  0.2923  0.2004 -0.2231  0.2954  0.3179  0.5523 -0.1615
#>   0.5823  0.1626  0.3005  0.1593 -0.3401 -0.3506  0.5533  0.1493 -0.3540
#>   0.3901  0.2709  0.1453  0.3591 -0.1540  0.7249  0.4389  0.5756 -0.2575
#> 
#> Columns 10 to 18 -0.2445  0.0191 -0.2150 -0.2623  0.1939 -0.0055  0.0214 -0.4466  0.0020
#>  -0.0969  0.0072 -0.0987 -0.5676  0.0011 -0.1564 -0.0302  0.2589  0.0173
#>  -0.3502 -0.3360 -0.4125 -0.1526  0.2324 -0.0570  0.0855 -0.2362 -0.1443
#> 
#> Columns 19 to 20  0.6366  0.3114
#>   0.6532  0.3685
#>   0.3906  0.2388
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.2316  0.5225  0.3737  0.4270 -0.0087  0.1217 -0.0775  0.4455 -0.0041
#>   0.5662  0.3807  0.2788  0.1048  0.1698  0.4111  0.5177  0.5274 -0.3335
#>   0.3202  0.2465 -0.1190  0.2331  0.0701  0.5641  0.0566  0.6532 -0.3748
#> 
#> Columns 10 to 18 -0.6223 -0.2385 -0.0549  0.0438 -0.0683 -0.1463  0.1849 -0.2988  0.3610
#>  -0.1563 -0.0976 -0.2970 -0.2574 -0.2072  0.2144  0.4466 -0.4032 -0.4420
#>  -0.4118 -0.4372 -0.5688 -0.0002  0.2910 -0.2781  0.0333 -0.4086 -0.1136
#> 
#> Columns 19 to 20  0.8981  0.1947
#>   0.6849  0.1646
#>   0.5804  0.0794
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.6250 -0.0010 -0.3375  0.2387 -0.1406 -0.4657 -0.1524 -0.1423 -0.4055
#>   0.1939 -0.4005 -0.0626  0.6261 -0.3476 -0.5310  0.4157 -0.4818  0.0772
#>   0.2413 -0.2399 -0.2318  0.7974  0.1746  0.7023 -0.4228 -0.3594  0.3984
#> 
#> Columns 10 to 18 -0.4352  0.1859  0.6323  0.1228  0.6392  0.5734 -0.5230  0.3222  0.3631
#>  -0.2409 -0.1854  0.1903  0.4159 -0.0925 -0.4377 -0.4426  0.4448 -0.4319
#>  -0.2298 -0.3988  0.2025 -0.1945 -0.3706 -0.4688 -0.2284 -0.3813 -0.0251
#> 
#> Columns 19 to 20  0.5020 -0.6082
#>  -0.3035 -0.0217
#>  -0.7046 -0.2713
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2316  0.5225  0.3737  0.4270 -0.0087  0.1217 -0.0775  0.4455 -0.0041
#>   0.5662  0.3807  0.2788  0.1048  0.1698  0.4111  0.5177  0.5274 -0.3335
#>   0.3202  0.2465 -0.1190  0.2331  0.0701  0.5641  0.0566  0.6532 -0.3748
#> 
#> Columns 10 to 18 -0.6223 -0.2385 -0.0549  0.0438 -0.0683 -0.1463  0.1849 -0.2988  0.3610
#>  -0.1563 -0.0976 -0.2970 -0.2574 -0.2072  0.2144  0.4466 -0.4032 -0.4420
#>  -0.4118 -0.4372 -0.5688 -0.0002  0.2910 -0.2781  0.0333 -0.4086 -0.1136
#> 
#> Columns 19 to 20  0.8981  0.1947
#>   0.6849  0.1646
#>   0.5804  0.0794
#> [ CPUFloatType{2,3,20} ]
#>