RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7724  0.8521  0.6519  0.3528  0.6170 -0.0306 -0.3484 -0.5138 -0.5458
#>   0.4752  0.3898  0.2566 -0.1155  0.0727 -0.4012  0.3426 -0.4574  0.2850
#>  -0.3904 -0.2825  0.3135  0.6732  0.3504  0.6329 -0.7223 -0.2668  0.9248
#> 
#> Columns 10 to 18  0.7947  0.8256  0.5734  0.4309 -0.2569 -0.3036 -0.6548  0.5631 -0.7754
#>   0.1897  0.7887  0.4046  0.4360  0.5906 -0.5030 -0.2083 -0.3300  0.5035
#>  -0.0191 -0.2069  0.7269 -0.1675 -0.6735 -0.0595 -0.9101 -0.4862 -0.9211
#> 
#> Columns 19 to 20  0.6601 -0.6965
#>   0.6558 -0.2729
#>  -0.4144  0.5680
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2309  0.2519  0.8963  0.7187 -0.3186 -0.0885 -0.4573 -0.5196 -0.2252
#>  -0.2108 -0.0669  0.2934  0.3727  0.1651  0.1184 -0.4621 -0.2432  0.0222
#>   0.3474  0.2654  0.0414  0.2185 -0.2929  0.2144 -0.1282  0.0556  0.3772
#> 
#> Columns 10 to 18  0.0878 -0.0919  0.5536  0.0629  0.2121 -0.4912 -0.2661 -0.6907  0.2108
#>   0.7624  0.6177  0.3566  0.4007  0.6065 -0.1909  0.3214 -0.7337 -0.2945
#>   0.0952  0.2208 -0.1345  0.6930  0.1438 -0.5679  0.4707 -0.3097 -0.0659
#> 
#> Columns 19 to 20  0.0224 -0.1998
#>   0.3759 -0.0009
#>  -0.0853 -0.3356
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1843 -0.1727  0.4877  0.1326 -0.0066  0.2344 -0.1725  0.1500 -0.1511
#>   0.2091  0.5716 -0.3189  0.4450  0.4967  0.1631  0.0101 -0.1538  0.0413
#>  -0.0042  0.4379 -0.4544 -0.4312  0.5643  0.1110 -0.0399 -0.0589  0.4233
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1040  0.3082  0.0705  0.0031 -0.6777  0.1081  0.1724 -0.0844 -0.2158
#>   0.1426  0.8531 -0.4627 -0.2876  0.2753 -0.1314 -0.0891  0.0703  0.2501
#>   0.3561 -0.4579  0.5688 -0.2700 -0.3668  0.5774 -0.6162  0.1490  0.0244
#> 
#> Columns 10 to 18 -0.1924 -0.3483 -0.1985 -0.1037 -0.1473 -0.5806 -0.6473 -0.5394 -0.0462
#>   0.3899  0.7983 -0.5795 -0.3178 -0.2031 -0.4384  0.5847 -0.2151  0.3799
#>   0.1663 -0.3421 -0.1961  0.2235  0.3527  0.6794  0.2283 -0.7519  0.1847
#> 
#> Columns 19 to 20  0.0186 -0.0734
#>   0.1983 -0.5866
#>   0.0340  0.7410
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0748  0.0927  0.5831  0.3636  0.2116  0.1675 -0.1671 -0.5187 -0.2985
#>   0.1014  0.1636  0.1465  0.2484  0.1920 -0.0434 -0.2017 -0.0991 -0.0440
#>   0.4508  0.2623  0.6865  0.2208 -0.2231  0.1309 -0.1583 -0.1735  0.0612
#> 
#> Columns 10 to 18  0.3269  0.1294  0.2889  0.1790  0.3343 -0.2621 -0.5466 -0.2279 -0.4864
#>   0.6127  0.4477  0.2912  0.2588  0.7424 -0.1596  0.5443 -0.7384 -0.2417
#>   0.5706  0.0242  0.2554  0.3078 -0.2537 -0.4747 -0.2999  0.0304 -0.3507
#> 
#> Columns 19 to 20  0.4542  0.0372
#>   0.3427 -0.2240
#>   0.0699 -0.1991
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>