RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6276  0.4542  0.7490 -0.1075  0.7961  0.3918 -0.6099 -0.8809 -0.4169
#>  -0.7103  0.3406  0.3135 -0.6608  0.8329 -0.4197  0.3708  0.4911  0.3665
#>  -0.8439  0.8099  0.3802  0.7878  0.8917 -0.4035  0.1950  0.9083 -0.3253
#> 
#> Columns 10 to 18 -0.4078  0.1961  0.0412  0.1174  0.3673 -0.1771 -0.2677 -0.0565  0.4600
#>   0.5994  0.2492 -0.6187 -0.4101  0.2504  0.3798  0.5351  0.0982  0.2259
#>  -0.1706 -0.0379 -0.1092 -0.3795  0.5742 -0.8376  0.3626  0.5003  0.5720
#> 
#> Columns 19 to 20  0.6557  0.6709
#>   0.8953 -0.5850
#>   0.2382 -0.8683
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0090 -0.2912 -0.3332 -0.4388  0.2338 -0.1580  0.0703 -0.4179 -0.7981
#>  -0.3320  0.2176  0.5594 -0.1143  0.6275  0.0203  0.4681  0.0833  0.3220
#>  -0.1848  0.4541  0.5241 -0.1341  0.1035 -0.1906 -0.0985  0.0553 -0.1233
#> 
#> Columns 10 to 18 -0.5294  0.3085  0.3654 -0.1988  0.2530  0.2435  0.0243 -0.5959  0.5625
#>  -0.0379 -0.4540 -0.0694  0.6586 -0.7187  0.0267  0.2012 -0.3324 -0.5475
#>  -0.3834  0.2782  0.2916  0.6364  0.3382 -0.4080  0.1988 -0.5779 -0.5143
#> 
#> Columns 19 to 20  0.5601 -0.5296
#>   0.4182  0.2684
#>  -0.3309 -0.4732
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4911  0.3033  0.1556 -0.4228  0.0393  0.4192  0.1186 -0.4861  0.1627
#>  -0.3940 -0.4941 -0.1966  0.2020  0.7903  0.1506  0.4858 -0.0983 -0.0885
#>  -0.4266 -0.1996  0.2285  0.1450  0.7611  0.0126 -0.0447  0.4909 -0.1945
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3485  0.2378 -0.2704  0.0416  0.3465 -0.3821 -0.5469  0.3163  0.4479
#>  -0.0519  0.0110  0.1486 -0.2104  0.4345  0.0915 -0.3230 -0.2443  0.0766
#>   0.4157  0.7423 -0.5441 -0.6247  0.2686 -0.1520 -0.0046 -0.4886  0.3434
#> 
#> Columns 10 to 18 -0.6593 -0.0452  0.6632  0.1294  0.1336 -0.2468  0.1845 -0.2041 -0.4884
#>  -0.5508 -0.6824  0.5640 -0.3637  0.1860 -0.0072  0.3400 -0.6946 -0.3175
#>   0.5564 -0.0984  0.2646  0.4897  0.3855  0.3157  0.4354  0.3966  0.4000
#> 
#> Columns 19 to 20 -0.3834 -0.1247
#>   0.2714 -0.9011
#>  -0.2898  0.6847
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5397  0.1152 -0.1306 -0.0251 -0.0707 -0.1005 -0.1495 -0.2086 -0.3158
#>  -0.1502 -0.0924  0.3693  0.2505  0.5807 -0.1932  0.4465  0.0854 -0.2577
#>   0.0710 -0.0499  0.0659  0.3665  0.7375 -0.3186  0.2389  0.3776 -0.2210
#> 
#> Columns 10 to 18 -0.1239  0.2335 -0.0103  0.0014 -0.1190  0.6353  0.1832 -0.6883  0.0640
#>   0.0243  0.1354  0.0039  0.1202  0.1333  0.1956  0.0752 -0.3314 -0.1879
#>  -0.2846 -0.0367  0.3362 -0.3146  0.4646 -0.5351 -0.1646 -0.2426 -0.3092
#> 
#> Columns 19 to 20  0.7048  0.2549
#>   0.1498 -0.0587
#>  -0.2394 -0.5052
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>