RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4944 -0.5454  0.0629  0.6060  0.2921  0.1756 -0.7672 -0.5683  0.2599
#>   0.3638 -0.3372  0.6386  0.6353 -0.7040  0.8281  0.4586  0.7437  0.5472
#>  -0.3633  0.0240 -0.8082  0.4048 -0.1717  0.8782  0.5281 -0.8897  0.3693
#> 
#> Columns 10 to 18  0.7468 -0.3282  0.2838 -0.2418  0.1112 -0.3472 -0.1618  0.6389 -0.8137
#>   0.4151 -0.6345  0.6357 -0.3028  0.7409 -0.3576 -0.7262  0.6527 -0.5119
#>  -0.4135 -0.3866 -0.5746  0.2435 -0.2158  0.4010  0.3047 -0.5537 -0.2320
#> 
#> Columns 19 to 20 -0.3002 -0.4557
#>  -0.4903  0.2910
#>  -0.4527  0.5851
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0404 -0.0042 -0.1285  0.0220 -0.3751  0.1680  0.3023 -0.0119 -0.0295
#>  -0.0609 -0.0173  0.1760 -0.0414 -0.0008  0.2139  0.1755  0.1389  0.3818
#>   0.1612 -0.2250 -0.0708  0.6899  0.4478  0.5965  0.4061 -0.5368  0.1062
#> 
#> Columns 10 to 18  0.6913 -0.0781  0.4509 -0.7990 -0.0794 -0.0025  0.0634 -0.0530 -0.5430
#>  -0.1230 -0.3596 -0.4047  0.1778  0.1694 -0.7973 -0.1393  0.2679 -0.6610
#>   0.0433 -0.2771  0.2210 -0.4741 -0.4965 -0.2567 -0.2084 -0.2571 -0.1637
#> 
#> Columns 19 to 20 -0.2566  0.2176
#>  -0.1974  0.2199
#>  -0.6082  0.1633
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2102 -0.3499  0.0570  0.2900 -0.0772  0.4890  0.6681  0.2748 -0.0241
#>  -0.1758 -0.1704 -0.0129  0.6321 -0.1166  0.3252  0.1147 -0.4103  0.4130
#>   0.5282 -0.2225  0.2434  0.6207  0.2296  0.3715  0.4344  0.0453 -0.0325
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6648  0.1854 -0.2190  0.0749 -0.4437 -0.5594 -0.0913  0.2794  0.0577
#>  -0.4686 -0.3968  0.5589 -0.3981 -0.4297 -0.1126 -0.4621 -0.0580  0.2063
#>   0.0786  0.4967 -0.0713 -0.8236 -0.5349 -0.1152 -0.6336 -0.5851  0.2424
#> 
#> Columns 10 to 18  0.3806  0.5660  0.7554  0.4378  0.5545  0.6297  0.0561 -0.4153 -0.6334
#>   0.1048  0.6692 -0.3245  0.7403 -0.2794  0.2642 -0.0715 -0.0544  0.6590
#>  -0.4540 -0.4870  0.1025 -0.6358  0.0844  0.4682 -0.4127  0.1055  0.5229
#> 
#> Columns 19 to 20 -0.5638 -0.6891
#>  -0.2055 -0.5912
#>  -0.2879  0.6919
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1837 -0.1461  0.2019  0.5808  0.5231  0.4339  0.1946 -0.2168  0.3051
#>   0.5023 -0.1916  0.4757  0.5900  0.0263  0.0696  0.2903  0.1399  0.0796
#>  -0.1618  0.2165  0.3548  0.3814 -0.2877  0.4902  0.5978  0.3264  0.1350
#> 
#> Columns 10 to 18  0.0345 -0.3557 -0.0550 -0.0590 -0.6058 -0.2635 -0.1913  0.1810 -0.3268
#>   0.1933 -0.2843  0.2122 -0.3570 -0.1886 -0.6337 -0.4303  0.0257 -0.0954
#>   0.3437 -0.4708  0.1884 -0.1162  0.2545 -0.2165 -0.0898 -0.2250 -0.7622
#> 
#> Columns 19 to 20  0.2082 -0.4109
#>   0.0571  0.0102
#>  -0.2637  0.0340
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>