RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6156  0.3968 -0.3266 -0.3300 -0.5322  0.1330 -0.0395  0.8991 -0.0753
#>  -0.7196 -0.7946  0.5622  0.6266 -0.1757  0.0460 -0.6798 -0.3759 -0.8775
#>  -0.8380 -0.1656  0.2089  0.4157 -0.8648 -0.3122  0.3780  0.6963  0.0127
#> 
#> Columns 10 to 18  0.6923 -0.4108 -0.1971  0.2855  0.5159  0.6041  0.7713  0.5219  0.5474
#>  -0.0246  0.3429 -0.5352 -0.7432 -0.3503  0.6717 -0.3975  0.5591 -0.3337
#>   0.5196  0.4356  0.3042 -0.6791  0.5650 -0.3308  0.5226  0.1735  0.3131
#> 
#> Columns 19 to 20 -0.0604 -0.9718
#>   0.0744  0.5371
#>   0.8324 -0.5235
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3332 -0.1980  0.2133  0.2953  0.0514  0.2359 -0.5941 -0.0646  0.2791
#>  -0.3220 -0.0796 -0.3430  0.1784  0.0781 -0.5917 -0.2058  0.6997  0.6957
#>  -0.3744 -0.0218  0.0597 -0.1641 -0.3705 -0.1881  0.3667  0.5084 -0.0379
#> 
#> Columns 10 to 18  0.1680 -0.4108 -0.2872 -0.6546 -0.4050  0.6579 -0.1582 -0.3723 -0.1885
#>   0.5544 -0.6342 -0.1314  0.0252 -0.1395 -0.0991  0.1406 -0.6877  0.4033
#>   0.2161  0.3758  0.1557 -0.0687 -0.0704  0.7810  0.6034  0.1241 -0.2067
#> 
#> Columns 19 to 20  0.0605 -0.2673
#>   0.0171 -0.6740
#>   0.5730 -0.4734
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3243 -0.0927 -0.3081  0.0936  0.4964 -0.1854 -0.2588  0.4000  0.4667
#>  -0.0497  0.2356 -0.1754  0.1415  0.3980  0.6585  0.5159 -0.3089 -0.3943
#>  -0.6828  0.0009 -0.0300  0.0318 -0.2981  0.2247 -0.2447  0.5147 -0.0799
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3649 -0.3159  0.4789  0.0134  0.8146  0.3362 -0.0144  0.2206  0.3355
#>   0.0873  0.3127 -0.4640  0.4215  0.2782  0.2064  0.7182 -0.1309  0.5941
#>  -0.4291  0.4400  0.8779  0.2937  0.5718  0.2429 -0.0936 -0.6216 -0.4981
#> 
#> Columns 10 to 18 -0.2655 -0.2590 -0.2824  0.5367  0.1814  0.7593  0.4080 -0.6072  0.1164
#>  -0.6476 -0.5622 -0.3410 -0.3186 -0.4711  0.6574 -0.2366 -0.5482  0.4910
#>  -0.7286  0.1837  0.7553  0.0108  0.2696  0.5909  0.2273 -0.5053 -0.5000
#> 
#> Columns 19 to 20 -0.1660 -0.0926
#>   0.3718 -0.2362
#>   0.4423  0.1187
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6070  0.2276 -0.4629 -0.0639  0.2106  0.0119 -0.3327  0.0958 -0.3394
#>  -0.2075 -0.0644 -0.3512  0.1140 -0.1101 -0.4484 -0.6413  0.1876  0.4357
#>  -0.5776  0.4104 -0.1179  0.0113  0.0948 -0.2822  0.2274  0.4204 -0.1184
#> 
#> Columns 10 to 18  0.3245 -0.2696 -0.4222 -0.3180  0.2204  0.1867  0.0188  0.2623  0.0624
#>   0.4048 -0.3571 -0.5271 -0.2253 -0.3813  0.5073 -0.2619 -0.1752  0.3258
#>   0.0185 -0.0313 -0.0305  0.1950 -0.0349  0.0345 -0.0176 -0.5267  0.0458
#> 
#> Columns 19 to 20 -0.2227 -0.1012
#>  -0.1858  0.3115
#>   0.2655  0.1068
#> [ CPUFloatType{2,3,20} ]
#>