RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2850  0.6574  0.0594 -0.1264  0.4896  0.1199 -0.7285 -0.1889  0.4723
#>  -0.2775 -0.7076 -0.0278  0.6727  0.5013 -0.3040 -0.1902 -0.5223  0.4214
#>  -0.6462 -0.6837 -0.4876  0.0152 -0.0505 -0.8051 -0.7245  0.0904 -0.4947
#> 
#> Columns 10 to 18  0.0602 -0.0313  0.4097  0.0355  0.1080 -0.1980 -0.3947 -0.9518  0.3864
#>   0.7605  0.1349 -0.5217 -0.5159 -0.3358  0.0017  0.2063 -0.7763 -0.1609
#>   0.5005  0.5663 -0.6602 -0.6048  0.4206  0.1964  0.0233 -0.5217 -0.5050
#> 
#> Columns 19 to 20 -0.6483  0.1354
#>  -0.6636  0.4123
#>  -0.3888 -0.5269
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0352  0.1848  0.2251 -0.1955  0.5080 -0.1481 -0.0869  0.7183  0.5375
#>  -0.2912  0.2777  0.0908 -0.3597  0.0257  0.2262  0.1143  0.5258 -0.1369
#>  -0.3419 -0.1495 -0.6001 -0.6905  0.1192  0.3953 -0.2858  0.4343 -0.4025
#> 
#> Columns 10 to 18  0.5167 -0.5317 -0.2218 -0.1299  0.7341 -0.1680 -0.2996  0.1984  0.6654
#>   0.4301  0.2412 -0.2248 -0.2847  0.5037 -0.3314 -0.6242 -0.0534  0.6394
#>   0.6078  0.4753 -0.5851 -0.1776  0.3804 -0.6529 -0.1993 -0.5873  0.7628
#> 
#> Columns 19 to 20  0.2209  0.1068
#>   0.3338 -0.0775
#>   0.5216 -0.4153
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5494 -0.3741  0.3869 -0.5956 -0.0115  0.3601 -0.4806  0.3075  0.4152
#>  -0.4129 -0.5053  0.2267 -0.4613  0.2214  0.3242 -0.3687 -0.2238  0.4670
#>  -0.1472  0.4833 -0.4325 -0.4505  0.1768 -0.5404 -0.4850  0.0010  0.5069
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0646 -0.1374 -0.5680  0.4757 -0.1327 -0.8472 -0.8408 -0.5864 -0.8265
#>   0.0688 -0.2683 -0.0293  0.6845 -0.4236  0.5857 -0.0195 -0.2026 -0.2651
#>   0.0583 -0.6526 -0.4454  0.2870 -0.6171  0.3368  0.2232  0.1812 -0.5058
#> 
#> Columns 10 to 18 -0.4592 -0.2132  0.1001  0.3509 -0.2704 -0.1110  0.1637  0.4127  0.1498
#>  -0.1842  0.2196  0.6767 -0.8828 -0.0464  0.3526  0.4633 -0.1941  0.3958
#>  -0.2474 -0.6772  0.6285  0.1682 -0.3921  0.4866 -0.2144 -0.2388 -0.2032
#> 
#> Columns 19 to 20 -0.2593 -0.3596
#>   0.6486  0.6422
#>   0.2503 -0.4796
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2747 -0.4218 -0.1320 -0.2983  0.2051  0.3749 -0.0774  0.5299  0.0450
#>  -0.0641 -0.1543 -0.0015 -0.5485  0.3785  0.1197  0.3571  0.0891  0.4682
#>  -0.3712 -0.3304  0.2027 -0.5008  0.1668 -0.1827  0.1714  0.1084  0.5006
#> 
#> Columns 10 to 18 -0.0599  0.1331  0.2350  0.1669 -0.1257 -0.0834 -0.0248 -0.0751  0.1416
#>   0.8481 -0.2316 -0.1602 -0.0583  0.4326 -0.0129 -0.4356 -0.2539  0.6051
#>   0.3317  0.0753  0.5868  0.4745  0.0873 -0.3052 -0.4237 -0.1860  0.2169
#> 
#> Columns 19 to 20 -0.0119 -0.4688
#>  -0.1914  0.0411
#>  -0.2963  0.1591
#> [ CPUFloatType{2,3,20} ]
#>