RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2811 -0.3661  0.6668 -0.5819  0.3723  0.5251  0.2342 -0.4042  0.0800
#>   0.6356  0.9558 -0.2900 -0.7009 -0.0318  0.1928 -0.4112  0.0659 -0.0176
#>  -0.5219  0.2717  0.5301  0.2731  0.5100  0.7720  0.4987 -0.2687  0.7160
#> 
#> Columns 10 to 18  0.6931 -0.4097 -0.7635 -0.7564  0.2186  0.3554 -0.5106  0.6224  0.3030
#>  -0.3203  0.1439  0.4072 -0.3490 -0.4214 -0.3646 -0.2882 -0.8901  0.6445
#>   0.4403 -0.5304 -0.7450  0.2019 -0.5977 -0.4269  0.0788  0.5642 -0.0554
#> 
#> Columns 19 to 20  0.6846 -0.2919
#>  -0.4992 -0.3792
#>  -0.5415  0.7651
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0653  0.2402  0.5798 -0.6887 -0.1289 -0.4774  0.3089  0.0262 -0.2646
#>  -0.0256  0.0260  0.2802 -0.0857  0.5812  0.2642 -0.3440 -0.2082  0.5659
#>   0.2445  0.5526  0.0091 -0.5961  0.2588 -0.4131  0.4964 -0.2139  0.2116
#> 
#> Columns 10 to 18  0.3890 -0.3757 -0.0986 -0.0946 -0.2272  0.2592  0.4590 -0.1212 -0.3840
#>   0.1071 -0.5742  0.2325  0.3345  0.0782  0.3642 -0.3476  0.4093  0.1229
#>   0.0079 -0.0599  0.5691 -0.0262 -0.5571  0.2741  0.3421 -0.4696  0.2218
#> 
#> Columns 19 to 20  0.3444  0.4737
#>  -0.0905  0.0624
#>   0.0994  0.7381
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1148  0.3575  0.5606 -0.2710  0.4084 -0.1695 -0.1652  0.1461  0.4880
#>   0.2583 -0.3396  0.1777 -0.2213  0.3225 -0.6024  0.0676 -0.0503 -0.0467
#>   0.0666  0.2133  0.4102 -0.6822  0.4735  0.1689  0.1439 -0.0809  0.4770
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5762  0.2737  0.8937  0.3400 -0.7119  0.7474  0.6975 -0.4622 -0.8544
#>   0.0585 -0.3040  0.4015  0.6752  0.1795  0.3681 -0.1779  0.5896  0.7602
#>   0.3285 -0.5993  0.3123 -0.0449  0.2940  0.1992  0.5465  0.4578 -0.2033
#> 
#> Columns 10 to 18  0.2980 -0.0436 -0.4357 -0.1063 -0.8363 -0.2337  0.5285 -0.4142 -0.7108
#>   0.5600  0.2652  0.6974  0.4694  0.5116  0.3341  0.7552 -0.0431  0.2455
#>   0.0290 -0.4333  0.0024 -0.4965 -0.2980 -0.5707 -0.1690  0.3316 -0.5774
#> 
#> Columns 19 to 20  0.6245 -0.1889
#>  -0.1967  0.8095
#>   0.3898 -0.5435
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2944  0.1457  0.3996 -0.3423 -0.2214 -0.7716 -0.0207 -0.0841  0.0375
#>  -0.1131 -0.1713  0.2462 -0.3917  0.4972 -0.0870  0.5786  0.0611  0.0775
#>  -0.2486  0.2200  0.3957  0.1410  0.3453 -0.3423 -0.0613 -0.1879  0.3196
#> 
#> Columns 10 to 18 -0.0462 -0.4411  0.4200 -0.0790 -0.4092  0.0166  0.5681 -0.6056  0.0493
#>   0.5243  0.1779 -0.3150  0.2135  0.3038  0.1686 -0.2898  0.1033 -0.2803
#>  -0.0907 -0.3988  0.5018  0.4366 -0.5704  0.5212  0.1952 -0.4177  0.0526
#> 
#> Columns 19 to 20 -0.0738  0.5595
#>   0.0593  0.1036
#>  -0.1342  0.3021
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>