RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2618 -0.2757  0.5824 -0.2520  0.4721 -0.3885  0.6518  0.3092 -0.3939
#>   0.5411 -0.1571 -0.2422  0.7333  0.3937 -0.6378 -0.1599 -0.9149  0.0681
#>  -0.8020 -0.5861 -0.2958 -0.4266  0.5632 -0.3444 -0.6115 -0.6457  0.6840
#> 
#> Columns 10 to 18  0.3715  0.4198  0.0559 -0.0214  0.2344 -0.6260 -0.2831  0.7872  0.2571
#>  -0.8533  0.6995  0.2181  0.1240  0.1864  0.8211  0.4575  0.8364  0.4786
#>  -0.9042  0.7906  0.1344  0.4496 -0.5034  0.8528 -0.4288  0.4805  0.2573
#> 
#> Columns 19 to 20  0.1757  0.3398
#>   0.6384  0.2755
#>   0.5532  0.6905
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2997 -0.5771  0.2620  0.0503  0.0638 -0.5930  0.0525  0.5759  0.1533
#>  -0.3108 -0.0724 -0.1911 -0.1329 -0.1196 -0.8144 -0.7442  0.4886 -0.3153
#>   0.0310  0.0686 -0.7420 -0.0690 -0.0169 -0.7134 -0.5501  0.0451 -0.3941
#> 
#> Columns 10 to 18 -0.3090  0.0183  0.2122  0.3914 -0.0468  0.2731 -0.1095  0.3979  0.1166
#>   0.2349 -0.1422 -0.0959 -0.4630 -0.6601  0.0804 -0.1454 -0.4868 -0.3327
#>   0.2687  0.0599 -0.0494 -0.6825 -0.3692 -0.5120 -0.1927 -0.7372 -0.7755
#> 
#> Columns 19 to 20  0.1312  0.7345
#>   0.0951  0.5025
#>  -0.0212  0.4671
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4869 -0.1975 -0.0417 -0.2443  0.1975 -0.6401 -0.0644  0.0357  0.3494
#>   0.2160 -0.5983 -0.3678 -0.0147  0.3384 -0.1333 -0.1704 -0.2128  0.2950
#>   0.5085 -0.3603  0.0591  0.1584  0.0580 -0.3172 -0.3340  0.4245  0.2077
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0342  0.3204  0.1798  0.7510  0.1719 -0.8706 -0.1757 -0.1850  0.7605
#>   0.5228  0.2821 -0.1381  0.0580 -0.0253 -0.5282  0.4927 -0.4221  0.1099
#>  -0.3036 -0.2908  0.0797 -0.6174  0.0601  0.0305  0.2902  0.4728  0.2777
#> 
#> Columns 10 to 18  0.1326 -0.0685  0.4813  0.0337  0.0844  0.4137  0.6552  0.4135  0.5384
#>  -0.1832  0.0298 -0.8270 -0.3500  0.5062  0.1617  0.5084 -0.3813  0.0355
#>  -0.0863 -0.7629 -0.4742 -0.3207  0.4942 -0.5224  0.2168  0.6647  0.5565
#> 
#> Columns 19 to 20 -0.6096  0.3668
#>  -0.0009 -0.0708
#>   0.5920  0.1324
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2201  0.3247  0.2025 -0.5765  0.2112 -0.4264 -0.1057  0.4823  0.3547
#>  -0.2426  0.1378 -0.3426  0.3333 -0.1244 -0.3768 -0.4851  0.0386  0.0066
#>   0.0739  0.3409 -0.5623  0.4188  0.1691 -0.6263 -0.5944  0.4217  0.0957
#> 
#> Columns 10 to 18  0.0668 -0.2873 -0.0786  0.1401 -0.0753 -0.3167  0.4202  0.1875  0.0314
#>  -0.2531 -0.2023  0.1917 -0.0774  0.2369  0.0085  0.5582  0.0411  0.0305
#>  -0.5276 -0.3492  0.0006 -0.0208  0.0756  0.1878  0.2727  0.0547 -0.1018
#> 
#> Columns 19 to 20 -0.4681  0.2753
#>  -0.2315  0.5300
#>  -0.1090  0.3059
#> [ CPUFloatType{2,3,20} ]
#>