RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8296  0.6952  0.0153  0.2471  0.3612 -0.9079  0.3701  0.4527 -0.4053
#>   0.6257  0.9650 -0.7056 -0.5856  0.3147 -0.7755 -0.3191  0.4010  0.4318
#>   0.2430  0.8550 -0.1140  0.2776 -0.6399 -0.5128 -0.0333  0.8560  0.0580
#> 
#> Columns 10 to 18 -0.0188  0.8217 -0.2426 -0.0989 -0.1873  0.6335  0.1907  0.1764  0.3471
#>  -0.0831 -0.2819 -0.6682 -0.2070 -0.5873 -0.2243  0.8969 -0.1872 -0.7550
#>   0.6838  0.5213  0.2720  0.0546  0.0398  0.0589 -0.1726  0.3461  0.0053
#> 
#> Columns 19 to 20  0.4642 -0.7340
#>   0.2088  0.6247
#>  -0.4069 -0.4067
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0513  0.2354  0.1418 -0.3780  0.1677  0.4185 -0.0254  0.6444 -0.0409
#>   0.2240  0.0319  0.1080 -0.3420  0.0015 -0.1113  0.6025  0.3749 -0.0848
#>  -0.1174  0.6654  0.2725 -0.2659  0.0714 -0.2538  0.4394  0.7224 -0.2370
#> 
#> Columns 10 to 18 -0.2039 -0.0891 -0.1888 -0.6905  0.3581  0.1372  0.8132 -0.5977 -0.5356
#>   0.1922  0.4375  0.0956  0.4615 -0.2812  0.5024  0.7077 -0.3773 -0.4893
#>   0.1727  0.1238  0.3086  0.0080 -0.1699  0.1719  0.4081 -0.4416 -0.2493
#> 
#> Columns 19 to 20 -0.0366  0.0033
#>   0.0879 -0.5930
#>   0.4401 -0.2081
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2836  0.1948 -0.0938  0.0466  0.0860  0.4598  0.0797  0.4547 -0.0094
#>  -0.0069  0.3328  0.2639  0.5080 -0.2368 -0.0213  0.0393  0.5402 -0.0182
#>   0.1277  0.3984  0.2014  0.5224  0.1562 -0.0376 -0.2329  0.5031 -0.2002
#> 
#> Columns 10 to 18 -0.5036 -0.1368 -0.5120 -0.0109 -0.0294 -0.0919  0.2991 -0.6023 -0.4022
#>  -0.4298 -0.1674 -0.0231  0.0133  0.1861 -0.1089  0.4934 -0.0405  0.1431
#>  -0.5492 -0.0694 -0.0316 -0.0916 -0.0306  0.2448  0.4331 -0.5348  0.3296
#> 
#> Columns 19 to 20  0.5688  0.2791
#>   0.5766  0.2125
#>   0.8236  0.2914
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.4454 -0.0290  0.3785  0.1479  0.1202  0.3742  0.1879  0.2232 -0.0708
#>   0.3504  0.0448  0.2755  0.3317  0.4333  0.3886  0.6462 -0.2731 -0.5890
#>   0.0538 -0.1380  0.4449  0.2207  0.5547  0.2002  0.2532 -0.3786 -0.3136
#> 
#> Columns 10 to 18 -0.4471 -0.1003 -0.4839  0.0335  0.1348 -0.0791 -0.0021 -0.3552  0.0296
#>   0.1222 -0.2058 -0.5662 -0.0058 -0.1790  0.2990  0.2023 -0.4890 -0.1244
#>  -0.2949 -0.0976 -0.3576 -0.1916 -0.2134  0.2101 -0.0252 -0.4800  0.2870
#> 
#> Columns 19 to 20  0.4733  0.1385
#>   0.4736 -0.3089
#>   0.6349  0.1230
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.3459 -0.0165  0.4499  0.4194  0.0172  0.1583  0.2510 -0.0644  0.1656
#>   0.2161  0.4255  0.3930  0.5644 -0.2270 -0.0209  0.1035  0.4851 -0.1837
#>  -0.0411 -0.3114  0.3971 -0.1758  0.1772  0.2575  0.1289  0.0819 -0.2230
#> 
#> Columns 10 to 18 -0.3867 -0.2499 -0.0588 -0.3246 -0.0551 -0.2947 -0.4403 -0.0178  0.0567
#>  -0.6503 -0.3939 -0.1441 -0.1456  0.2483 -0.0357  0.1287  0.2720  0.2457
#>  -0.3657 -0.2130 -0.5027 -0.6067 -0.0254 -0.3400  0.0271 -0.0724 -0.0803
#> 
#> Columns 19 to 20  0.4772 -0.0412
#>   0.7332  0.5848
#>   0.5213  0.1467
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2753  0.3697  0.0979 -0.4934  0.1064 -0.1332 -0.0181  0.1744 -0.6996
#>  -0.2376  0.8285 -0.2144  0.1474 -0.4406 -0.5108 -0.6929 -0.4945  0.5025
#>   0.0567 -0.1559 -0.1636 -0.3208 -0.6816  0.6861 -0.6522 -0.5077 -0.2978
#> 
#> Columns 10 to 18  0.2724 -0.3142 -0.3097 -0.4453  0.3344 -0.1425  0.3273  0.3617 -0.1309
#>   0.2253  0.3319  0.0613 -0.3550  0.1609  0.6140 -0.0859  0.0476 -0.2900
#>   0.3681 -0.3431  0.1367 -0.1914  0.1133 -0.4043 -0.0050 -0.4591  0.6889
#> 
#> Columns 19 to 20  0.1786 -0.7074
#>   0.4243 -0.4765
#>  -0.1280 -0.1215
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3459 -0.0165  0.4499  0.4194  0.0172  0.1583  0.2510 -0.0644  0.1656
#>   0.2161  0.4255  0.3930  0.5644 -0.2270 -0.0209  0.1035  0.4851 -0.1837
#>  -0.0411 -0.3114  0.3971 -0.1758  0.1772  0.2575  0.1289  0.0819 -0.2230
#> 
#> Columns 10 to 18 -0.3867 -0.2499 -0.0588 -0.3246 -0.0551 -0.2947 -0.4403 -0.0178  0.0567
#>  -0.6503 -0.3939 -0.1441 -0.1456  0.2483 -0.0357  0.1287  0.2720  0.2457
#>  -0.3657 -0.2130 -0.5027 -0.6067 -0.0254 -0.3400  0.0271 -0.0724 -0.0803
#> 
#> Columns 19 to 20  0.4772 -0.0412
#>   0.7332  0.5848
#>   0.5213  0.1467
#> [ CPUFloatType{2,3,20} ]
#>