RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2552  0.4305  0.9288  0.0970 -0.1112 -0.4004 -0.8833 -0.1168 -0.3329
#>   0.0554 -0.3290  0.8435 -0.6070 -0.0291  0.0412 -0.0886 -0.3579 -0.6916
#>   0.5866  0.0275 -0.8185  0.2273 -0.5807 -0.2542 -0.1372 -0.3361 -0.2483
#> 
#> Columns 10 to 18  0.3922 -0.6530 -0.1246  0.0657  0.2650 -0.8360 -0.0041  0.2933  0.4218
#>   0.3998 -0.9255 -0.6752 -0.0915  0.6796  0.2077 -0.4614 -0.7013 -0.6637
#>   0.6283  0.9029 -0.4098 -0.2374  0.5701  0.5198  0.1669  0.4034  0.5585
#> 
#> Columns 19 to 20 -0.4433 -0.2607
#>   0.5925 -0.5766
#>   0.0543 -0.2622
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3754  0.1330  0.3273 -0.2832  0.5125 -0.2822 -0.1011 -0.1048 -0.4510
#>  -0.2916  0.3574  0.4648 -0.4857  0.8660 -0.1662 -0.5006 -0.4531 -0.0336
#>   0.5738 -0.0937  0.0088  0.3853  0.1953  0.4804 -0.4195 -0.3228  0.5387
#> 
#> Columns 10 to 18 -0.3829  0.2507 -0.4037  0.1432 -0.2441 -0.2296  0.1164  0.6519 -0.5325
#>  -0.0742  0.1534 -0.5160 -0.2542 -0.2854 -0.8702  0.2867  0.3245  0.2561
#>  -0.1522  0.3862 -0.4134 -0.3335  0.3462 -0.4989 -0.5249  0.2335 -0.2183
#> 
#> Columns 19 to 20  0.4751 -0.4292
#>   0.0480 -0.4115
#>  -0.3214 -0.2839
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2979  0.6031 -0.3525 -0.0302  0.8551 -0.4966 -0.0420 -0.2280 -0.1313
#>   0.4362  0.6569 -0.2079  0.0970  0.7309  0.0702 -0.6160 -0.0192 -0.1686
#>   0.3592  0.5739  0.3281  0.0478  0.5528 -0.3707 -0.4257 -0.0954 -0.1754
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5760 -0.5771 -0.5868 -0.5524 -0.1332  0.4537  0.5010 -0.0604 -0.1466
#>   0.3090  0.7363 -0.4718 -0.1410 -0.0619  0.8021  0.3731  0.8863 -0.7751
#>  -0.0177 -0.5766 -0.1252 -0.6891 -0.1095  0.3035 -0.0854  0.0575  0.2313
#> 
#> Columns 10 to 18  0.4176 -0.4364 -0.0494  0.1694 -0.4658  0.4233 -0.6844 -0.1510 -0.1188
#>   0.2065  0.1909 -0.4499 -0.2136  0.4277  0.4861  0.1570 -0.0228  0.6073
#>   0.0491  0.4248  0.2334  0.4378 -0.4047  0.2051 -0.4736 -0.1114  0.0138
#> 
#> Columns 19 to 20  0.3270 -0.1234
#>   0.5056 -0.2909
#>   0.2660 -0.3592
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0943  0.0346 -0.2114 -0.2459  0.7393 -0.5698 -0.3488 -0.4859 -0.1972
#>   0.6824  0.4557  0.2363  0.2569  0.5653  0.1802 -0.2644 -0.0374 -0.1328
#>   0.0672  0.1991 -0.1197 -0.2291  0.5959 -0.1924 -0.0946 -0.2562 -0.2654
#> 
#> Columns 10 to 18 -0.2274  0.2799 -0.4182 -0.1975  0.0068 -0.3241 -0.5205  0.5965 -0.5138
#>   0.3013  0.5689 -0.6229 -0.3985  0.5805 -0.4534 -0.6181  0.5280 -0.3230
#>  -0.3749  0.1154 -0.0809 -0.1842 -0.0131 -0.3025 -0.6135  0.3829 -0.3059
#> 
#> Columns 19 to 20  0.1686 -0.4739
#>  -0.0846 -0.3921
#>  -0.2747 -0.3622
#> [ CPUFloatType{2,3,20} ]
#>