RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5189 -0.1630  0.4691 -0.0105  0.1078  0.3014  0.0412  0.8012 -0.0044
#>  -0.6872  0.6357  0.5379  0.3246  0.8532 -0.1795  0.2448 -0.5645 -0.8572
#>   0.9189 -0.2065 -0.3344  0.1849 -0.8890 -0.0271  0.5968  0.5878  0.4599
#> 
#> Columns 10 to 18 -0.7117  0.3480  0.3242  0.7493 -0.9004  0.3484  0.5189 -0.3604  0.1897
#>  -0.4880 -0.2446  0.6107  0.8090 -0.8754 -0.7755  0.3246 -0.6202  0.4372
#>   0.7920  0.5351  0.2716  0.6680 -0.2069  0.9193 -0.3312 -0.5617 -0.2112
#> 
#> Columns 19 to 20 -0.5789 -0.0797
#>  -0.0265  0.2703
#>   0.1217  0.3959
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0474 -0.2058  0.0700 -0.6214 -0.0626  0.3011  0.1194 -0.1765  0.3381
#>  -0.4224 -0.3248  0.1959 -0.5455  0.7729  0.2867  0.1347  0.4572  0.6118
#>   0.4668  0.4577  0.0682 -0.2125  0.6945 -0.0832 -0.4255 -0.7760 -0.6207
#> 
#> Columns 10 to 18 -0.5112 -0.5565  0.1313  0.2248  0.1386  0.2723 -0.3529 -0.2335  0.1129
#>  -0.3304 -0.6577 -0.1487  0.3955 -0.4367  0.1301 -0.3774  0.3509  0.4922
#>  -0.5123 -0.3669  0.2143 -0.2343 -0.0736 -0.1046 -0.5415 -0.0589  0.0939
#> 
#> Columns 19 to 20 -0.5082 -0.0659
#>   0.2611  0.3895
#>   0.2859 -0.3765
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5500  0.0720 -0.5154  0.0722 -0.1211  0.2713  0.3119  0.1005 -0.0615
#>  -0.6817 -0.5418 -0.3000 -0.3669  0.0076  0.3749  0.4310  0.4277  0.3823
#>   0.0291 -0.1453 -0.1392 -0.0363 -0.0938 -0.3824  0.3227  0.2041  0.1357
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1026  0.3261  0.3908  0.1269  0.0143  0.1826 -0.0530 -0.1314 -0.3872
#>  -0.7155  0.6593  0.6456 -0.6313  0.1830  0.5902  0.2864  0.8615 -0.6377
#>  -0.3277  0.3854  0.7394 -0.4561 -0.2916  0.4318 -0.6519  0.8529 -0.4435
#> 
#> Columns 10 to 18  0.2093  0.0042 -0.3379  0.0808 -0.8323  0.1787  0.2380 -0.0861 -0.3438
#>   0.6802 -0.1278 -0.4346 -0.6924 -0.8570 -0.4127  0.4093 -0.6946 -0.1741
#>   0.6262 -0.4209 -0.2636 -0.4820 -0.5131 -0.4603 -0.0939 -0.1033  0.3417
#> 
#> Columns 19 to 20  0.4259 -0.0866
#>   0.0028  0.8130
#>   0.0134  0.7169
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2467 -0.1423 -0.1608 -0.3708 -0.0142  0.0380  0.1652  0.2337  0.5733
#>  -0.2684 -0.3508 -0.4407 -0.6997 -0.2564  0.5063  0.4632  0.4029  0.6029
#>   0.1678 -0.0454 -0.2497 -0.6009 -0.0584  0.5028  0.5999  0.1962  0.4800
#> 
#> Columns 10 to 18 -0.3845  0.1033  0.3032 -0.2936 -0.1620 -0.0927 -0.2532  0.3864  0.1108
#>  -0.0808 -0.3517  0.5750 -0.1646 -0.2299 -0.0699 -0.4214  0.3597  0.0569
#>  -0.3759 -0.4176  0.3878  0.0819  0.0619 -0.2077 -0.1414  0.5041  0.1566
#> 
#> Columns 19 to 20 -0.2801  0.1004
#>  -0.4491 -0.3488
#>   0.1089 -0.2534
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>