RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.9214 -0.2675  0.8649  0.4498  0.7816  0.3304  0.6033  0.6395 -0.4837
#>  -0.4781 -0.1879 -0.7438 -0.1448 -0.0617 -0.1536  0.4893 -0.1264  0.3687
#>   0.0601  0.6715  0.2678 -0.1609  0.3027  0.7478  0.5190  0.6306  0.5309
#> 
#> Columns 10 to 18 -0.6983  0.2893 -0.2143  0.2917  0.7666  0.8138  0.5231 -0.4322  0.9255
#>  -0.7426  0.0722  0.3055  0.1394 -0.3867 -0.3892 -0.1337  0.5228  0.5553
#>   0.5620  0.5681 -0.8308  0.6086 -0.6553 -0.2274 -0.1852  0.5354 -0.3480
#> 
#> Columns 19 to 20  0.7703 -0.1874
#>  -0.4629  0.1872
#>  -0.4590 -0.5262
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3392 -0.7856  0.1664  0.6285 -0.0186 -0.0318 -0.3889  0.0234 -0.1152
#>  -0.3638  0.6518  0.1581 -0.3378  0.5099 -0.4808  0.3926 -0.2957 -0.3254
#>  -0.5827 -0.5960  0.0838 -0.6721  0.2407  0.0226 -0.3179  0.1475 -0.2173
#> 
#> Columns 10 to 18 -0.4897 -0.1862 -0.3379 -0.4581 -0.6181  0.0246  0.0782  0.5802  0.4636
#>   0.1495  0.1777  0.2962 -0.3489  0.1384  0.4082  0.3273 -0.4639 -0.0707
#>  -0.5243  0.3384 -0.6346 -0.3969  0.3562  0.7183  0.3204 -0.3875 -0.1853
#> 
#> Columns 19 to 20  0.6457 -0.3111
#>  -0.3446  0.5166
#>  -0.4260  0.1708
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4463  0.3240  0.4835 -0.1718  0.5162 -0.6191  0.3579 -0.2395 -0.0225
#>  -0.2401  0.4970 -0.1459 -0.0135  0.2246  0.2224  0.7101 -0.3886  0.1399
#>  -0.2811  0.6256  0.1840 -0.1920  0.1106 -0.2014  0.5524  0.2419  0.5480
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0674  0.3199 -0.2798 -0.0126  0.4922 -0.1099 -0.2407  0.0290  0.5368
#>  -0.1087  0.1791 -0.1836 -0.2611  0.1102  0.3021  0.4321  0.2117  0.5093
#>   0.1507  0.1874 -0.0485 -0.4401  0.3241  0.3045 -0.1746  0.4796  0.2491
#> 
#> Columns 10 to 18 -0.0654 -0.2657 -0.2234  0.3394 -0.4344 -0.5017  0.0570  0.2229  0.2468
#>  -0.3204 -0.4884  0.5104  0.2055  0.1280  0.0341 -0.3106  0.5146  0.6432
#>   0.1934 -0.3381  0.7157  0.3629  0.1975 -0.0127 -0.3144  0.4361  0.1904
#> 
#> Columns 19 to 20 -0.1730  0.3288
#>   0.1877  0.1001
#>  -0.0328  0.0142
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2183  0.3869  0.2368  0.4049  0.1819 -0.3367  0.4221 -0.4278 -0.1366
#>  -0.1709  0.4444 -0.2096 -0.1211  0.2315 -0.2124  0.2750  0.2162 -0.0506
#>  -0.1951  0.4617  0.3306 -0.1205  0.4873 -0.3755  0.4722 -0.1150 -0.0396
#> 
#> Columns 10 to 18 -0.6329  0.1239  0.1356  0.0578 -0.5122 -0.1245  0.3036  0.4116  0.4225
#>  -0.3354 -0.2493  0.0293  0.0973 -0.2250  0.4824 -0.0423  0.2174  0.1707
#>  -0.4638  0.0433  0.1376 -0.0954 -0.1784  0.4495  0.4316 -0.0469  0.0859
#> 
#> Columns 19 to 20 -0.1755 -0.0215
#>  -0.5936 -0.3847
#>  -0.3113 -0.0648
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>