RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6876 -0.0405 -0.7708  0.5750  0.4794 -0.0509 -0.8094 -0.7211  0.0399
#>   0.3401 -0.5826 -0.6833  0.5032 -0.5242 -0.2586 -0.1960  0.8252  0.2483
#>  -0.2595 -0.2421 -0.1009 -0.0538 -0.2278  0.6311 -0.3719 -0.2042 -0.0480
#> 
#> Columns 10 to 18 -0.6218  0.6564 -0.7405 -0.4334  0.2944  0.8998  0.3955  0.9131 -0.0064
#>  -0.8308 -0.0580  0.2456 -0.3352  0.6134  0.6095 -0.5228 -0.1878 -0.2822
#>   0.5951  0.1732 -0.3900 -0.5926  0.5515 -0.2451  0.3111 -0.1541  0.4682
#> 
#> Columns 19 to 20  0.0388 -0.9168
#>  -0.6235  0.0000
#>  -0.6411 -0.6069
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6977 -0.0809 -0.5609  0.2938 -0.5534  0.2301 -0.1641  0.0281 -0.5267
#>  -0.1738  0.0894 -0.3360  0.3006 -0.2163 -0.1133 -0.6965  0.1339 -0.6656
#>   0.1230 -0.0800 -0.0648  0.4349 -0.4790 -0.2773 -0.0485  0.2705 -0.1414
#> 
#> Columns 10 to 18  0.4567 -0.2134 -0.7890 -0.2301  0.2641  0.1128 -0.5866 -0.1972 -0.3701
#>   0.1152  0.3923 -0.6670 -0.1434  0.1289  0.4893  0.7452 -0.1805 -0.4109
#>   0.6582 -0.4576 -0.5660 -0.3109  0.1160  0.7210 -0.3363 -0.1411 -0.5292
#> 
#> Columns 19 to 20 -0.6800  0.6279
#>  -0.1668 -0.3519
#>   0.1848  0.1034
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2342  0.0620 -0.0660  0.2202 -0.5286 -0.3462 -0.6578  0.0474 -0.7875
#>   0.0722  0.0891 -0.0648  0.3976 -0.1100  0.1882 -0.0197 -0.1241 -0.7553
#>   0.0270 -0.0063  0.2508  0.2802 -0.0567 -0.0207 -0.2995  0.2959 -0.5019
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1021  0.3604 -0.5498  0.1893 -0.4975  0.1313  0.5101  0.6259 -0.1566
#>  -0.3445  0.6730  0.3725 -0.1576  0.4534 -0.0618 -0.2068 -0.0840 -0.3183
#>   0.2014  0.7690  0.0464 -0.3537 -0.1510  0.2756  0.4124  0.5927 -0.4088
#> 
#> Columns 10 to 18 -0.1113 -0.2713 -0.1491  0.2100  0.6005  0.5275 -0.3047  0.2035 -0.2106
#>   0.6014 -0.6980  0.0960 -0.1825  0.1605 -0.3612 -0.1544  0.1995 -0.0751
#>   0.0005  0.0521  0.4458 -0.1853  0.5881  0.1605 -0.0096 -0.1973  0.3915
#> 
#> Columns 19 to 20  0.1486  0.5802
#>   0.5131 -0.6034
#>   0.2850  0.3036
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0774 -0.0194 -0.1767  0.1199 -0.3869 -0.1885 -0.4677  0.3264 -0.8590
#>   0.5352  0.3805  0.0638  0.2774  0.0046 -0.3854  0.0973  0.0277 -0.1207
#>   0.2335  0.1631  0.0041  0.0342 -0.2463 -0.3353 -0.5148 -0.0232 -0.5601
#> 
#> Columns 10 to 18  0.2769 -0.2576 -0.6912 -0.1896  0.0363  0.3937  0.1924 -0.3800 -0.4341
#>  -0.0579 -0.2928 -0.7444 -0.1062 -0.1881  0.5946 -0.1396  0.1956 -0.6296
#>   0.1660 -0.0317 -0.7369 -0.2705  0.3064  0.5357  0.1542 -0.0125 -0.4551
#> 
#> Columns 19 to 20 -0.4064  0.1694
#>  -0.1701  0.5641
#>  -0.6008 -0.1766
#> [ CPUFloatType{2,3,20} ]
#>