RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6260 -0.4627  0.4035  0.6639 -0.1405 -0.1136 -0.7625 -0.7949 -0.7138
#>  -0.3601  0.4206  0.9187  0.6688 -0.3300  0.1916  0.2925  0.2807  0.0575
#>  -0.4375  0.6782  0.2241 -0.6829  0.8800  0.7655 -0.2870  0.1234  0.2537
#> 
#> Columns 10 to 18  0.5169 -0.4495  0.1485 -0.1022 -0.6970 -0.0900 -0.0739  0.9075  0.5587
#>  -0.6493  0.0852  0.7786  0.4242  0.6078  0.3242  0.5888  0.7851 -0.1773
#>   0.6706 -0.3006  0.5764 -0.9639 -0.6759 -0.7061  0.2425 -0.3235  0.4507
#> 
#> Columns 19 to 20 -0.3622 -0.7654
#>   0.3383  0.1318
#>   0.5410  0.7683
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4534 -0.2673 -0.1652  0.0529 -0.1842  0.1037  0.3416 -0.3138 -0.3982
#>  -0.4761 -0.2347  0.4460  0.5021  0.4243 -0.2766  0.0839 -0.1845  0.3008
#>   0.0901 -0.3498  0.1501  0.0296 -0.7145 -0.2271  0.6937 -0.3854 -0.7535
#> 
#> Columns 10 to 18  0.3258  0.0063 -0.3347  0.2572  0.5492 -0.2588 -0.0551  0.2614  0.3215
#>  -0.3010  0.5430 -0.0474  0.0611 -0.6296 -0.2012  0.3042  0.6053 -0.1062
#>   0.3890  0.5134 -0.1462 -0.1827  0.6454 -0.0521 -0.6272 -0.4469  0.1869
#> 
#> Columns 19 to 20 -0.0851  0.3834
#>  -0.5190 -0.3626
#>   0.3087 -0.7464
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0876 -0.1486  0.4523  0.0795  0.3063 -0.4265 -0.3655  0.4873 -0.3201
#>  -0.4851 -0.0879  0.0515 -0.1358  0.3346  0.3132 -0.0398 -0.3820 -0.2985
#>  -0.3813  0.0934  0.1388  0.0881  0.7519  0.5332 -0.4400  0.3869  0.3660
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7078 -0.3643 -0.1549  0.2064  0.5546 -0.8228  0.3945  0.3244 -0.1513
#>  -0.4155  0.3819  0.2631 -0.2933 -0.8183  0.6975  0.1148  0.3814  0.0296
#>   0.0323  0.0292  0.4961  0.3816  0.7936  0.3751 -0.5880 -0.7037 -0.2075
#> 
#> Columns 10 to 18 -0.6642 -0.2393  0.6143 -0.0164  0.2130  0.8754  0.1723  0.6202 -0.4789
#>  -0.5446 -0.0809 -0.5823  0.0253 -0.8352 -0.3241 -0.4295  0.7026  0.6031
#>   0.4018 -0.5547 -0.4321  0.1085 -0.0463 -0.6363  0.3448 -0.8074  0.3362
#> 
#> Columns 19 to 20 -0.5341 -0.3830
#>   0.5044  0.4621
#>  -0.4716 -0.4709
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2991  0.2736  0.5628 -0.4359 -0.0165  0.2953  0.1447  0.3577 -0.4003
#>  -0.5380  0.0558 -0.2446  0.2257 -0.1305 -0.2688 -0.4709 -0.2875 -0.5879
#>  -0.4781  0.2040 -0.0155  0.3583  0.6546  0.3701  0.2421  0.1237  0.0919
#> 
#> Columns 10 to 18  0.3517  0.2123  0.4793 -0.4277  0.0270 -0.3212 -0.4152  0.3202  0.2608
#>  -0.0941  0.2762 -0.3100  0.2206 -0.6206 -0.3844 -0.3632 -0.0785  0.0323
#>   0.2486  0.2382  0.5776 -0.3238 -0.2200 -0.5106  0.1078 -0.2049  0.2950
#> 
#> Columns 19 to 20  0.3285  0.0958
#>   0.1483 -0.6104
#>  -0.1545  0.2836
#> [ CPUFloatType{2,3,20} ]
#>