RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3530 -0.1082  0.6245  0.6273  0.7717  0.4488 -0.6541 -0.1812 -0.3904
#>   0.2029  0.0254  0.6947 -0.3646  0.2995 -0.3183  0.3372 -0.7493  0.1027
#>  -0.2266 -0.2830  0.6482  0.8476  0.2573 -0.1912  0.5373 -0.3355  0.5977
#> 
#> Columns 10 to 18  0.6076  0.7412  0.0013  0.5535  0.3878 -0.5836 -0.3290 -0.7666 -0.7184
#>   0.2937  0.3568 -0.4549 -0.5396  0.9071  0.0680 -0.5374  0.6134  0.6161
#>  -0.0764  0.8149  0.5477  0.1505  0.4787 -0.5267 -0.1741 -0.1468 -0.2983
#> 
#> Columns 19 to 20  0.0870 -0.5020
#>  -0.0443 -0.6828
#>   0.0373  0.6954
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5135  0.2257 -0.0508  0.3111 -0.7226  0.1380  0.1289 -0.2768  0.0097
#>  -0.0864  0.3881 -0.0987  0.0571  0.3350 -0.0716 -0.2811  0.1605  0.5432
#>  -0.8429  0.1469 -0.4844  0.4561 -0.3065  0.0566  0.0316 -0.1079  0.4499
#> 
#> Columns 10 to 18  0.2045 -0.0742 -0.1593  0.1273  0.8029 -0.4115 -0.4728  0.6954 -0.6828
#>   0.0386  0.2729 -0.7073  0.1156  0.3592 -0.0761  0.0469  0.3326  0.3908
#>  -0.7233 -0.0048  0.0725  0.0489  0.8035  0.1608 -0.3654  0.1681 -0.0208
#> 
#> Columns 19 to 20 -0.3717  0.0921
#>   0.1592  0.0033
#>   0.2930  0.5477
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5962  0.1269 -0.6180  0.0238  0.0636 -0.0788 -0.2980 -0.0979  0.5687
#>  -0.1037  0.1808 -0.3618  0.1543 -0.2642 -0.3414 -0.1994 -0.2280  0.0394
#>  -0.6117  0.1244 -0.4703  0.4918 -0.0410 -0.3119 -0.3490 -0.0107 -0.1341
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4235  0.1023  0.1322 -0.6903  0.4342 -0.3921 -0.3444 -0.8013  0.2576
#>   0.4613 -0.0929 -0.3193  0.1805  0.0053  0.0017  0.1170 -0.4117  0.2302
#>  -0.7199 -0.5681  0.6315  0.7184  0.1862 -0.1309  0.2466  0.4588 -0.2370
#> 
#> Columns 10 to 18  0.1772 -0.7903 -0.5972 -0.2091  0.5431 -0.1088 -0.2427 -0.6541  0.2526
#>   0.1909  0.3717  0.2168 -0.0714  0.4627 -0.2739 -0.5020 -0.6874  0.7002
#>   0.1616 -0.4796 -0.2485  0.6781 -0.3142  0.5237 -0.1344  0.1920 -0.0961
#> 
#> Columns 19 to 20 -0.0788  0.8054
#>   0.6166  0.3844
#>  -0.1195 -0.3324
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5485  0.0039 -0.1861 -0.1510 -0.5144 -0.1849  0.3431  0.1153  0.1433
#>  -0.3744 -0.4739  0.0066  0.1609 -0.4130 -0.0290 -0.0771 -0.3344  0.1832
#>  -0.2859 -0.0431 -0.3499  0.1548 -0.4885  0.1191 -0.4244  0.0163  0.4032
#> 
#> Columns 10 to 18 -0.3620  0.0362  0.1268 -0.2176  0.4276 -0.1513 -0.0526  0.5063  0.2086
#>  -0.1987  0.0283  0.0377  0.0186  0.3918  0.2081  0.0947  0.5002 -0.0588
#>   0.2053  0.0658 -0.5507 -0.1154 -0.2262 -0.3749 -0.3273  0.0525 -0.0369
#> 
#> Columns 19 to 20  0.0290 -0.5354
#>   0.2939 -0.0137
#>  -0.2401  0.3850
#> [ CPUFloatType{2,3,20} ]
#>