RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3445  0.2915 -0.3985  0.4592  0.9115  0.8716  0.4991 -0.1201  0.2711
#>  -0.2394 -0.5700  0.0990 -0.7676 -0.0032  0.2626  0.4143 -0.0422  0.8621
#>  -0.1929 -0.1819 -0.0588 -0.4486  0.1151  0.6629  0.7461 -0.2683  0.5107
#> 
#> Columns 10 to 18 -0.7578 -0.5580  0.8390 -0.7724  0.0203 -0.1633 -0.2656 -0.3732 -0.1157
#>   0.9234 -0.6640 -0.8868  0.7445  0.8558  0.1133 -0.0580  0.4232 -0.6058
#>   0.6990 -0.7248 -0.2695  0.2282  0.4121  0.6414  0.0835 -0.6539 -0.4470
#> 
#> Columns 19 to 20 -0.3296  0.8424
#>  -0.0182 -0.1119
#>  -0.5080 -0.3536
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0880 -0.3878  0.6312 -0.1220  0.5665  0.2364  0.5635  0.1898  0.0191
#>   0.3019  0.2765  0.3108 -0.2151 -0.1826  0.0557 -0.4854 -0.3488  0.3739
#>   0.1239 -0.3657  0.5241  0.0956  0.0621  0.3038 -0.1855  0.1870  0.2604
#> 
#> Columns 10 to 18  0.6383 -0.2975 -0.5311  0.4040  0.1859 -0.0061 -0.0662 -0.0237 -0.2230
#>   0.3886 -0.4321  0.3676  0.1032 -0.5645 -0.0171 -0.8471 -0.5617 -0.0177
#>   0.2921 -0.1226 -0.3059  0.6676 -0.2715  0.1676 -0.8848 -0.7100  0.3536
#> 
#> Columns 19 to 20 -0.5497  0.8105
#>  -0.2955  0.4347
#>  -0.3726  0.5621
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0062 -0.2483  0.2712 -0.2353  0.6638  0.2247 -0.1000  0.1950  0.3215
#>  -0.7280 -0.2777 -0.2114 -0.4857 -0.2881  0.1173  0.0491  0.5115  0.4498
#>  -0.4054  0.0422 -0.2196 -0.2086  0.0757  0.1899  0.1521 -0.0163  0.3600
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5753 -0.0198 -0.4746 -0.4190 -0.2322 -0.2044  0.5269 -0.0907 -0.3220
#>  -0.2725  0.2310  0.4878  0.2479 -0.1227  0.1078  0.2210 -0.4946  0.2570
#>   0.1291  0.5079  0.6346 -0.2237  0.3855 -0.0188 -0.2144 -0.2783  0.4261
#> 
#> Columns 10 to 18 -0.5359 -0.4354  0.1704 -0.2233 -0.1643 -0.5895  0.0729  0.7600 -0.2586
#>   0.0275  0.1569  0.1280  0.0337 -0.6643  0.1759  0.4855  0.1720  0.1463
#>  -0.3189  0.4310  0.6468 -0.1952 -0.8601 -0.1045  0.0345  0.3554 -0.4480
#> 
#> Columns 19 to 20  0.7862 -0.4306
#>   0.0054 -0.2523
#>  -0.2660  0.7820
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0176  0.2654  0.1757 -0.7252 -0.2576  0.0228 -0.1011 -0.0928  0.4396
#>  -0.0189 -0.3996  0.0055 -0.5265 -0.0531  0.1746  0.0769  0.1041  0.2209
#>  -0.1253 -0.2582 -0.2927 -0.5072  0.0118  0.2654  0.3164  0.2113 -0.1900
#> 
#> Columns 10 to 18  0.6433 -0.3373 -0.1259  0.3993  0.0127 -0.3597 -0.5261  0.1871 -0.2431
#>   0.0767 -0.0504 -0.1999  0.5264  0.4053  0.0011 -0.5085  0.0138 -0.2168
#>   0.3868 -0.3118 -0.1238  0.2887  0.3815 -0.1567 -0.5233 -0.1994 -0.1553
#> 
#> Columns 19 to 20 -0.0150 -0.0409
#>   0.1398  0.4846
#>  -0.0742  0.2636
#> [ CPUFloatType{2,3,20} ]
#>