RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3135 -0.2814  0.2983 -0.4735  0.2045  0.1065  0.3609  0.3710 -0.7071
#>  -0.2226  0.1947 -0.1217 -0.6857  0.6345 -0.3362 -0.0064  0.1464 -0.0367
#>   0.2907 -0.7745 -0.6751 -0.5325  0.3810  0.8760  0.3600  0.5376 -0.0617
#> 
#> Columns 10 to 18 -0.2444  0.0034  0.1153  0.5956  0.2440  0.8810  0.7836 -0.8761 -0.2836
#>   0.5925 -0.4887 -0.1586  0.4234  0.0253  0.9078 -0.0036 -0.1481  0.5451
#>  -0.3434  0.7863 -0.3540 -0.0564  0.0005 -0.3439  0.3339  0.7971  0.0370
#> 
#> Columns 19 to 20 -0.0093 -0.1679
#>   0.0188 -0.0108
#>   0.2317  0.1686
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0384  0.4738 -0.1039  0.0079 -0.0577 -0.0492  0.1448  0.4803 -0.6001
#>  -0.2261  0.1404 -0.2724 -0.1241 -0.0686  0.2145  0.0952  0.2941 -0.2008
#>   0.3162 -0.1593  0.0973 -0.6006 -0.3574  0.3757  0.2307  0.4754 -0.2905
#> 
#> Columns 10 to 18 -0.1511  0.2795  0.6591 -0.5468 -0.2992  0.0499  0.3114 -0.2285 -0.4543
#>  -0.1874  0.3671  0.3513 -0.5240  0.5407 -0.2454 -0.0854 -0.1149  0.4630
#>   0.1271  0.3195  0.2841 -0.4224  0.2259 -0.0355  0.3624 -0.4435 -0.2046
#> 
#> Columns 19 to 20 -0.6515  0.4459
#>   0.1666  0.5680
#>  -0.1094  0.2847
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0067  0.5763  0.1064 -0.1983 -0.1329  0.4841  0.2267  0.6342 -0.5007
#>   0.0589  0.1768  0.1013 -0.0754  0.3079 -0.0094 -0.0823  0.1342 -0.0269
#>  -0.1066 -0.2330  0.1812 -0.4501  0.5022  0.0628  0.6100  0.2728 -0.0703
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3253 -0.1314  0.3381 -0.3890 -0.0828 -0.3734 -0.4646 -0.5062 -0.2789
#>   0.2015 -0.3194  0.5012 -0.1681  0.3628 -0.2498 -0.8138 -0.6956 -0.3749
#>  -0.2671 -0.7044  0.4773 -0.1187  0.6822 -0.8295 -0.2173 -0.0796 -0.5471
#> 
#> Columns 10 to 18 -0.2195  0.2995 -0.1915  0.3914 -0.4395 -0.5738  0.2771 -0.6943 -0.6023
#>   0.4247  0.6306 -0.0225 -0.1358 -0.6225  0.1601 -0.0016 -0.7631 -0.0716
#>  -0.1632  0.6293  0.5050 -0.1246  0.0033  0.7725  0.2438 -0.0259 -0.0738
#> 
#> Columns 19 to 20  0.7378 -0.3192
#>   0.6741  0.2591
#>   0.0168 -0.1885
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0518  0.3392 -0.3210 -0.1099  0.0514  0.6381  0.2912  0.3768 -0.2467
#>   0.2503  0.4441 -0.0594 -0.4406 -0.0480  0.6028  0.6136  0.4806 -0.5575
#>  -0.2326  0.3853  0.2008 -0.3203  0.1093  0.2926  0.5670  0.4466 -0.3573
#> 
#> Columns 10 to 18 -0.1107  0.3483  0.1589 -0.0962 -0.1386  0.1966  0.0138  0.1181 -0.3940
#>  -0.1949  0.6959  0.2369 -0.3469  0.0730 -0.0858  0.3742  0.2174  0.0081
#>  -0.1642  0.0583  0.1294 -0.1131  0.4729 -0.0293  0.0613 -0.1530 -0.2650
#> 
#> Columns 19 to 20 -0.3581 -0.0017
#>   0.1006 -0.2282
#>  -0.0122 -0.1042
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>