RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2521 -0.7193 -0.3578  0.6003  0.6665 -0.7984 -0.5970 -0.6103  0.9219
#>   0.5472 -0.1683 -0.1710  0.2499 -0.3593 -0.7019  0.4129 -0.1054  0.3155
#>   0.0950  0.1535 -0.3045 -0.2167 -0.4033  0.8896 -0.0332 -0.4186  0.7482
#> 
#> Columns 10 to 18  0.4863 -0.4711  0.7004 -0.8181 -0.4020 -0.4964 -0.2736  0.8616 -0.3145
#>  -0.0191  0.6394 -0.3493 -0.0089 -0.3398  0.3428  0.4440 -0.4189 -0.2809
#>  -0.4818 -0.2003  0.6389 -0.7245  0.6694 -0.0091 -0.8027 -0.4242 -0.9727
#> 
#> Columns 19 to 20  0.3586 -0.2484
#>   0.1402  0.5564
#>  -0.4304  0.6065
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.8146  0.8078 -0.2919 -0.0367  0.2439 -0.6732 -0.4673  0.1989 -0.0269
#>   0.7374  0.0932  0.1334 -0.1376 -0.1279  0.1561 -0.1679  0.3763  0.5414
#>   0.6675  0.2740  0.1113  0.0575 -0.6480 -0.1278 -0.1843 -0.4377 -0.5448
#> 
#> Columns 10 to 18  0.7349 -0.5234  0.1745  0.0808  0.4161 -0.4685 -0.2743  0.0275 -0.1025
#>   0.3723 -0.0868 -0.4057  0.0700  0.1444 -0.4929  0.1133 -0.2197 -0.4890
#>   0.4757 -0.3422  0.1651  0.3043  0.4060  0.0573 -0.4164 -0.1077  0.1431
#> 
#> Columns 19 to 20 -0.6140  0.1258
#>  -0.4537  0.2340
#>  -0.3320  0.5202
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.7853  0.1691  0.1233 -0.0318 -0.2120 -0.5536  0.5391  0.4590 -0.2256
#>   0.7614  0.1153 -0.3494  0.1703 -0.3654 -0.3191  0.0861  0.6425  0.4302
#>   0.6554  0.1730  0.1951 -0.2050 -0.0838 -0.4311  0.3643  0.3433 -0.1663
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1767 -0.4175  0.2677  0.6671  0.1967  0.7706 -0.5572  0.1392  0.3689
#>   0.0007  0.0706 -0.1012  0.7484  0.6760  0.1823 -0.5588 -0.2823  0.6935
#>   0.8723 -0.4753 -0.5597 -0.2784  0.7411  0.6647 -0.1075  0.6399  0.8266
#> 
#> Columns 10 to 18 -0.3113 -0.6651  0.6046  0.8860  0.1548  0.6401  0.0553  0.4967 -0.4283
#>   0.3369 -0.5198 -0.3114  0.8515 -0.1703 -0.0106  0.3604  0.8528 -0.2494
#>  -0.3216  0.4420  0.3329 -0.0950 -0.6493 -0.5340 -0.2827  0.2817 -0.0639
#> 
#> Columns 19 to 20  0.0094  0.2832
#>  -0.1745 -0.1581
#>   0.3727 -0.0271
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4909 -0.3503  0.2267  0.1557 -0.1101 -0.0091  0.4815  0.2081  0.1287
#>   0.4905  0.0315  0.0026  0.2477 -0.1075 -0.3705  0.2893  0.4254  0.2769
#>   0.7289 -0.2659 -0.3304  0.2988 -0.8470 -0.3114  0.0845 -0.2709 -0.0508
#> 
#> Columns 10 to 18  0.3264  0.0386 -0.2151  0.3935  0.5266 -0.1114 -0.2381  0.0625  0.1429
#>   0.5842 -0.0423  0.1606  0.3462  0.1697 -0.2286 -0.3132 -0.2211 -0.4014
#>   0.4083  0.0595 -0.5105  0.3520 -0.1027  0.0766 -0.6354 -0.2151 -0.1511
#> 
#> Columns 19 to 20 -0.4710 -0.0448
#>  -0.1018  0.4111
#>  -0.0974 -0.1698
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>