RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8186 -0.3696 -0.7929 -0.0696 -0.1393  0.4102 -0.5464 -0.1939 -0.2442
#>  -0.3296 -0.0873 -0.2021 -0.4084 -0.6017  0.3241 -0.3793  0.6821  0.3043
#>   0.7012  0.3336 -0.2203  0.1502 -0.1941  0.1043 -0.5724 -0.1415  0.5279
#> 
#> Columns 10 to 18  0.4717  0.3063  0.8590  0.9515  0.3303  0.4879  0.1284  0.4079 -0.6322
#>  -0.6050  0.7583  0.8424  0.6658  0.7866 -0.2993  0.6394  0.9951 -0.7808
#>  -0.2507  0.8099  0.3932 -0.2613 -0.5657  0.6152 -0.0831 -0.4697 -0.6671
#> 
#> Columns 19 to 20  0.3036  0.9195
#>   0.5055  0.7009
#>  -0.9459 -0.4186
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1316 -0.2513  0.0519 -0.8067  0.4120  0.1661 -0.2605 -0.3340 -0.2335
#>  -0.0834 -0.1686  0.2389 -0.4487  0.3167 -0.3962 -0.4282 -0.5705 -0.2085
#>  -0.3457  0.1607 -0.3905  0.0458 -0.5850 -0.0178 -0.4224  0.4926  0.3650
#> 
#> Columns 10 to 18  0.0686  0.2995  0.4649 -0.0690 -0.2556  0.2685 -0.0287  0.0911 -0.6016
#>   0.4042 -0.1585 -0.4566 -0.1157 -0.1841  0.2201 -0.1104 -0.4176 -0.2349
#>  -0.3072  0.3634  0.4591  0.0447 -0.4969  0.1943 -0.1024  0.5259 -0.2412
#> 
#> Columns 19 to 20 -0.7891  0.2032
#>  -0.5032 -0.8420
#>  -0.4361 -0.1307
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4395 -0.3366  0.0832 -0.2237  0.4182 -0.2068 -0.6611 -0.5552 -0.0438
#>  -0.2571  0.4262 -0.2574  0.2267  0.1821  0.0504 -0.5335  0.3921  0.0182
#>  -0.3903 -0.4408  0.2924 -0.3492  0.3166 -0.0276 -0.4195 -0.1290 -0.1927
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2016 -0.2597 -0.3854  0.1866 -0.3110 -0.3080 -0.7456  0.0193 -0.5221
#>   0.4214 -0.2550 -0.7650  0.1591 -0.1921  0.1086  0.2338  0.2293 -0.1192
#>  -0.1574 -0.1392  0.3399  0.0970  0.0633 -0.0488  0.3215  0.5826  0.1359
#> 
#> Columns 10 to 18  0.4043  0.5353 -0.4172  0.6298 -0.1913  0.1651  0.6943  0.4135  0.2145
#>   0.1654  0.8081 -0.5161 -0.0448 -0.0618  0.5625  0.0595  0.4440 -0.2204
#>   0.4907  0.0049 -0.3923  0.7338  0.2803 -0.3896  0.0582  0.3619 -0.2506
#> 
#> Columns 19 to 20  0.3992  0.0631
#>   0.6068  0.1833
#>  -0.5681 -0.3903
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2147 -0.5092 -0.1579 -0.3652  0.3213  0.0415 -0.2503 -0.2262  0.3004
#>  -0.2989 -0.0442 -0.0174 -0.4116  0.5118 -0.1059 -0.4562 -0.0101 -0.0293
#>  -0.3222 -0.7185 -0.2116 -0.4966  0.3923  0.1553 -0.5003 -0.1276 -0.2114
#> 
#> Columns 10 to 18 -0.0547  0.0334  0.1266 -0.0086 -0.2027  0.6807 -0.2374  0.0243 -0.1478
#>   0.1808  0.0673  0.3771  0.0067 -0.3085  0.3393 -0.1100  0.0975 -0.2193
#>   0.1069  0.1334 -0.0274  0.0586 -0.0078  0.5204  0.2407  0.1129 -0.5797
#> 
#> Columns 19 to 20 -0.5037  0.6000
#>  -0.3933  0.1036
#>  -0.3945  0.6614
#> [ CPUFloatType{2,3,20} ]
#>