RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1231  0.0188  0.1713  0.0769 -0.0583 -0.1851 -0.2678 -0.4169  0.1469
#>  -0.7858 -0.2293 -0.7687 -0.1631  0.0415  0.7212 -0.6099 -0.3243 -0.2917
#>   0.2135 -0.5103  0.2712 -0.0952 -0.3681  0.1184 -0.6626 -0.7235 -0.0541
#> 
#> Columns 10 to 18 -0.1518 -0.8300  0.3203  0.2450 -0.4037  0.0937 -0.2534  0.5182 -0.1277
#>   0.4960 -0.5242  0.1281  0.5455  0.1609 -0.2790  0.6757  0.0963  0.0137
#>  -0.8086  0.4186  0.6312 -0.1826 -0.2434 -0.1796 -0.4409 -0.4203 -0.4594
#> 
#> Columns 19 to 20  0.1850 -0.5452
#>   0.0016 -0.1062
#>   0.5317  0.1168
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1461 -0.1406  0.3286  0.1613  0.3018  0.1223 -0.0900 -0.4611 -0.0198
#>  -0.0716 -0.0689  0.4046  0.2588 -0.5437  0.6369  0.0232 -0.5427  0.4488
#>  -0.1043  0.1581  0.1758  0.4001 -0.1184  0.1791 -0.3635 -0.2807 -0.0244
#> 
#> Columns 10 to 18 -0.5708 -0.0702  0.0649 -0.1908  0.3460 -0.2337  0.0283  0.0498 -0.6600
#>  -0.8173 -0.2879 -0.0493  0.1117  0.2992 -0.5757  0.5798 -0.1261 -0.4513
#>   0.0061 -0.3239  0.4121 -0.3355 -0.0958  0.1850 -0.0424 -0.5008 -0.4596
#> 
#> Columns 19 to 20  0.1388  0.2576
#>  -0.1020  0.3425
#>  -0.1278  0.3764
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0053 -0.0930  0.2400  0.6648  0.1937  0.5362 -0.2358 -0.2408 -0.1356
#>  -0.0051  0.1592 -0.4519  0.0141  0.1243  0.5361 -0.2336 -0.1598  0.2793
#>  -0.1249  0.1225  0.3368  0.6132 -0.1390  0.3275 -0.1716  0.0682  0.1773
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5007  0.3867  0.2427 -0.2891  0.0522 -0.4132 -0.7294 -0.1443  0.4081
#>  -0.0469 -0.1521 -0.6765 -0.2798  0.6459  0.0787 -0.3984 -0.4846  0.1341
#>  -0.3659  0.1275 -0.6745  0.2199  0.3803  0.1017 -0.5508  0.1132 -0.5813
#> 
#> Columns 10 to 18 -0.3605  0.4232  0.2212 -0.5514  0.6800  0.0837  0.4463 -0.6791 -0.5255
#>  -0.4150  0.7436 -0.7290 -0.4965 -0.4425 -0.2871 -0.4359  0.4768  0.5733
#>  -0.1026  0.3037 -0.1105 -0.5784 -0.4731  0.0424 -0.1525  0.1725 -0.3738
#> 
#> Columns 19 to 20 -0.6975 -0.4021
#>  -0.0318  0.1881
#>  -0.2087  0.1755
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0152  0.2613 -0.0232  0.4726 -0.0445  0.4572 -0.5721 -0.3489  0.7359
#>  -0.3458  0.3596  0.0613 -0.0430  0.0212  0.3539 -0.0640  0.4712 -0.4351
#>  -0.5260  0.2827 -0.0444  0.2597 -0.0245  0.5053 -0.0805 -0.2814  0.1773
#> 
#> Columns 10 to 18  0.3643  0.6982  0.4541 -0.2131 -0.4134  0.5202 -0.3163  0.4333 -0.1725
#>  -0.0082 -0.2847  0.1551 -0.3508 -0.3577 -0.0931 -0.1927  0.0404 -0.0868
#>  -0.2027 -0.1985  0.0393 -0.2146 -0.0543  0.1947 -0.1032  0.0587 -0.4009
#> 
#> Columns 19 to 20  0.4345  0.0875
#>   0.4036  0.3288
#>   0.2304  0.0635
#> [ CPUFloatType{2,3,20} ]
#>