RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.0774  0.1615  0.1870 -0.2400 -0.5199  0.3932  0.0067  0.4831  0.5752
#>  -0.3467  0.6772  0.3317  0.6656 -0.4203  0.6454  0.4641  0.9414  0.5318
#>   0.4249  0.9453  0.7278 -0.1167 -0.5728 -0.3657 -0.0589  0.7849 -0.6837
#> 
#> Columns 10 to 18 -0.7943 -0.5052 -0.2296 -0.1850  0.0857 -0.2685  0.6106  0.6032 -0.8235
#>   0.8016  0.5834 -0.1424  0.5599 -0.3484  0.1803  0.3012  0.2151 -0.2343
#>   0.5851 -0.1329  0.4313 -0.4016 -0.0207 -0.1681 -0.4167  0.6613 -0.5008
#> 
#> Columns 19 to 20  0.5964  0.8094
#>   0.0670 -0.6153
#>   0.2649 -0.5733
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2925  0.3552  0.3329  0.6437  0.0592  0.5262  0.0607  0.6667  0.0315
#>   0.1908  0.9027  0.5029  0.4845 -0.2112 -0.2078  0.1494  0.7200 -0.3670
#>   0.5551  0.2100  0.5777  0.3336 -0.4090  0.1445  0.3664  0.7340 -0.6428
#> 
#> Columns 10 to 18 -0.1349  0.1149 -0.3598  0.2250  0.4491  0.1432  0.1601 -0.1343  0.1367
#>  -0.2574 -0.0592 -0.3839  0.2187  0.5166  0.0646  0.2747 -0.4680  0.5726
#>  -0.0919 -0.1294  0.0129  0.1494  0.4764  0.4988 -0.2540 -0.5561 -0.1013
#> 
#> Columns 19 to 20  0.6543 -0.1828
#>   0.6119  0.3799
#>   0.4701 -0.4719
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3592  0.5444  0.4515  0.4643  0.1295  0.2863  0.4187  0.0316  0.2716
#>  -0.0001 -0.1808  0.4836  0.5861  0.3215  0.5846  0.2947 -0.2100 -0.3312
#>   0.1574  0.0673  0.4959  0.3365 -0.4598  0.4879  0.0433  0.6834 -0.1371
#> 
#> Columns 10 to 18 -0.4204 -0.2206 -0.1485 -0.0703 -0.0456  0.0471  0.0419 -0.2976 -0.1974
#>  -0.3561 -0.5238 -0.4376  0.1534  0.2647  0.1129 -0.1582 -0.4939  0.2807
#>  -0.4392 -0.6060 -0.0989  0.0825  0.3449 -0.3094 -0.1290 -0.2977 -0.0270
#> 
#> Columns 19 to 20  0.5415  0.1995
#>   0.4778 -0.3141
#>   0.4118 -0.1037
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.4773  0.1483  0.5334  0.3610 -0.3572  0.3940  0.3300  0.6718 -0.1087
#>   0.1534 -0.1342  0.2961  0.0927 -0.2410  0.5221  0.3279  0.3201  0.1152
#>   0.6287  0.3486  0.2537  0.4832 -0.1332  0.4728  0.5939  0.3253 -0.1977
#> 
#> Columns 10 to 18 -0.2562  0.1323 -0.2447  0.0266  0.5822  0.1099 -0.0180 -0.3033  0.1617
#>  -0.3558 -0.3098 -0.4092 -0.6202  0.1978 -0.5279 -0.2285  0.1632 -0.2874
#>  -0.0830 -0.1875 -0.3319  0.0426  0.0490 -0.3458 -0.2160 -0.4152 -0.0974
#> 
#> Columns 19 to 20  0.6835  0.1782
#>   0.4613  0.2201
#>   0.6640  0.1761
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.3017  0.4574  0.3942  0.5431  0.1703  0.4366 -0.0352  0.5129 -0.0946
#>   0.6721  0.4746  0.1917  0.5114 -0.3147 -0.0128  0.6854  0.2082 -0.4461
#>   0.4917  0.5053  0.5013  0.4003 -0.1390  0.1297  0.4446  0.1512 -0.1373
#> 
#> Columns 10 to 18 -0.5747 -0.3032 -0.3975  0.3107  0.2915 -0.0799  0.2860 -0.5049  0.2972
#>  -0.0910  0.0725 -0.2106 -0.0173 -0.0194 -0.1730 -0.1579 -0.0649  0.2090
#>  -0.5830 -0.1148  0.0204  0.0910  0.0095 -0.0379 -0.1835 -0.1922  0.3052
#> 
#> Columns 19 to 20  0.8015  0.2573
#>   0.7316 -0.1101
#>   0.8337  0.2509
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.4875 -0.3792 -0.6719  0.6580  0.0926 -0.0469  0.0257 -0.3765  0.3169
#>   0.0688  0.3137  0.0455 -0.0801 -0.3507 -0.5485  0.0299  0.4810  0.0583
#>   0.6535 -0.0679 -0.0210  0.2163 -0.2514 -0.3926  0.3766 -0.2362 -0.4421
#> 
#> Columns 10 to 18 -0.0175  0.4655  0.3847 -0.0344 -0.0709  0.1993 -0.6498 -0.2641  0.1302
#>   0.4690  0.3341  0.1727  0.6413 -0.1401  0.5613 -0.3612  0.5179 -0.4724
#>  -0.2655  0.1451 -0.0429  0.1676  0.3674  0.3224 -0.4242  0.5669 -0.2387
#> 
#> Columns 19 to 20  0.3040  0.2097
#>  -0.5166 -0.2417
#>   0.4194 -0.5265
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3017  0.4574  0.3942  0.5431  0.1703  0.4366 -0.0352  0.5129 -0.0946
#>   0.6721  0.4746  0.1917  0.5114 -0.3147 -0.0128  0.6854  0.2082 -0.4461
#>   0.4917  0.5053  0.5013  0.4003 -0.1390  0.1297  0.4446  0.1512 -0.1373
#> 
#> Columns 10 to 18 -0.5747 -0.3032 -0.3975  0.3107  0.2915 -0.0799  0.2860 -0.5049  0.2972
#>  -0.0910  0.0725 -0.2106 -0.0173 -0.0194 -0.1730 -0.1579 -0.0649  0.2090
#>  -0.5830 -0.1148  0.0204  0.0910  0.0095 -0.0379 -0.1835 -0.1922  0.3052
#> 
#> Columns 19 to 20  0.8015  0.2573
#>   0.7316 -0.1101
#>   0.8337  0.2509
#> [ CPUFloatType{2,3,20} ]
#>