RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4016 -0.1163  0.3772 -0.8554 -0.0560  0.1176 -0.3883  0.8197  0.2193
#>  -0.5214 -0.3471 -0.1117  0.4982  0.6393  0.3212  0.3539 -0.0033 -0.1450
#>   0.1640  0.8361  0.1044  0.6708  0.5091  0.8726 -0.4974  0.3514 -0.3831
#> 
#> Columns 10 to 18  0.6885 -0.0540 -0.6045 -0.1970  0.4558 -0.0412  0.8235 -0.5767 -0.9590
#>  -0.3435  0.3246 -0.0394  0.2161 -0.6072  0.4093  0.4120 -0.5966 -0.2483
#>  -0.6330 -0.4686 -0.3946 -0.5788  0.2076 -0.1244 -0.2666 -0.7793  0.6808
#> 
#> Columns 19 to 20 -0.2906  0.3112
#>   0.2337 -0.3152
#>   0.6623  0.7684
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4498  0.2810 -0.5068  0.1496  0.1058  0.1550  0.1234  0.2056  0.0491
#>   0.0940 -0.2070  0.5630 -0.1168  0.0679  0.0952  0.5812 -0.1001 -0.2692
#>   0.6577  0.0263 -0.2810  0.0935  0.3318  0.2256  0.2686 -0.3322 -0.6398
#> 
#> Columns 10 to 18 -0.0274 -0.2325 -0.4570  0.3679 -0.1061 -0.3868  0.2606 -0.5742 -0.0679
#>  -0.3332 -0.1951 -0.3223 -0.4002  0.3291 -0.2173  0.2503 -0.1587 -0.1187
#>   0.0252 -0.1379 -0.6921  0.1391 -0.0483 -0.3079 -0.3984 -0.8636  0.1597
#> 
#> Columns 19 to 20  0.1336  0.1098
#>  -0.1151  0.2811
#>   0.7458 -0.1375
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1147  0.0864  0.2677  0.0300  0.0506  0.1859  0.5685  0.2939 -0.6564
#>   0.4274  0.3087  0.0869  0.5233  0.2784  0.4151 -0.0116  0.4533 -0.0626
#>   0.4461  0.0087  0.4821 -0.2775 -0.1501  0.2190  0.3812  0.5463 -0.2092
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6454  0.4219 -0.2710  0.5066  0.5126 -0.3505  0.7223 -0.1692 -0.2265
#>  -0.1190  0.5623 -0.7577  0.0379 -0.1206  0.3051 -0.6677 -0.5786  0.2499
#>   0.0420 -0.0404 -0.2299  0.2845  0.3683  0.5742 -0.1670  0.0275 -0.4100
#> 
#> Columns 10 to 18  0.4229  0.3467 -0.5608  0.4225 -0.9102 -0.3456 -0.5498  0.7129 -0.7825
#>   0.6219 -0.1829 -0.1905 -0.1133  0.5802  0.4233 -0.6327 -0.3907 -0.1786
#>  -0.5808 -0.6719 -0.3316  0.3033 -0.5406 -0.6330 -0.1111 -0.5070 -0.0690
#> 
#> Columns 19 to 20 -0.5493 -0.0490
#>  -0.4026  0.0004
#>  -0.0670 -0.5109
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6112  0.0861  0.2009  0.5656 -0.2115  0.4948  0.6764  0.1116 -0.2231
#>   0.1590  0.0464  0.6044  0.5818 -0.4573  0.2834  0.4158  0.5870 -0.2680
#>   0.3245  0.1960  0.3044  0.1391 -0.2267  0.6122  0.0394  0.5448  0.1785
#> 
#> Columns 10 to 18  0.0447 -0.3834 -0.2359  0.2562  0.1950 -0.0358  0.0274 -0.3838 -0.3876
#>  -0.4778 -0.2885 -0.2762  0.1610  0.4512 -0.1119 -0.0494  0.0299  0.0589
#>  -0.4189 -0.2234 -0.3830 -0.2415  0.3136 -0.3381 -0.2372 -0.2506 -0.0071
#> 
#> Columns 19 to 20  0.3507 -0.0423
#>   0.5950  0.0750
#>   0.5569  0.2264
#> [ CPUFloatType{2,3,20} ]
#>