RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9191 -0.4099  0.4806 -0.1896  0.1973 -0.3483  0.8789  0.3833 -0.2599
#>  -0.3603 -0.0383 -0.0954  0.4047  0.1642  0.4775  0.2388  0.2253  0.1051
#>  -0.1163  0.8743  0.0072 -0.2724  0.7437  0.8071  0.0046 -0.5375  0.0541
#> 
#> Columns 10 to 18  0.9614 -0.4540 -0.7364 -0.3997 -0.2782  0.6113  0.8846  0.6366  0.6155
#>   0.7669  0.4318 -0.4559 -0.4332 -0.4116  0.0899  0.4482 -0.3118  0.2526
#>  -0.8261  0.5267 -0.5305  0.5485  0.3529  0.0201 -0.7359 -0.2298  0.2133
#> 
#> Columns 19 to 20 -0.7723  0.8479
#>   0.2596 -0.2945
#>  -0.2063  0.1489
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4170  0.3398  0.2414 -0.4061  0.7253 -0.0633  0.6014 -0.5703 -0.4332
#>   0.0039 -0.4806  0.3361 -0.1463  0.2917  0.1426  0.1209  0.2576 -0.4271
#>  -0.0927 -0.1461 -0.0701  0.2492 -0.2916 -0.0653  0.4863  0.2058  0.0151
#> 
#> Columns 10 to 18 -0.1543  0.6408 -0.1537  0.4308  0.0584 -0.0081  0.1418  0.3559  0.3626
#>   0.1689  0.2873 -0.3210 -0.1801 -0.1787 -0.2555  0.4356 -0.5676  0.2506
#>  -0.1639  0.3858 -0.6945 -0.0394 -0.3469  0.2049  0.2374 -0.4413  0.2622
#> 
#> Columns 19 to 20  0.1471  0.2966
#>   0.2592  0.3957
#>  -0.5948 -0.0754
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4463  0.0317  0.0273  0.1477 -0.1584 -0.5838  0.3563  0.2178 -0.1322
#>  -0.0782 -0.1802  0.1747 -0.1538  0.3139 -0.1240  0.2143  0.1581 -0.0152
#>  -0.3952  0.5179 -0.1941 -0.1154  0.4310 -0.0535  0.3626  0.0430 -0.2348
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2545  0.1653 -0.2453 -0.5194  0.7840  0.8835  0.3105 -0.2117 -0.3768
#>   0.5397  0.5797  0.0164  0.2897 -0.4186  0.2492  0.0741  0.5306 -0.2138
#>  -0.2679 -0.6128 -0.0310  0.1306 -0.3026  0.1679 -0.1961  0.5604 -0.5465
#> 
#> Columns 10 to 18  0.8096 -0.3424 -0.1540 -0.6693 -0.8386  0.9320  0.4448 -0.4229 -0.0554
#>   0.2518  0.0300 -0.8456  0.3112 -0.1314  0.3196 -0.4201  0.3333 -0.7197
#>  -0.0010  0.0907 -0.4386  0.0253 -0.4796  0.1489 -0.2520  0.3505  0.3639
#> 
#> Columns 19 to 20  0.2859 -0.0639
#>  -0.4898  0.2208
#>  -0.4123 -0.6912
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1594  0.4704 -0.5440  0.1469  0.6872  0.1080 -0.3738 -0.0887 -0.3550
#>  -0.5208 -0.5295 -0.0954  0.0640  0.1432 -0.3247  0.4141 -0.0383  0.2651
#>   0.1098  0.4401  0.4067 -0.0320  0.1941 -0.2779  0.5280  0.0557 -0.2210
#> 
#> Columns 10 to 18 -0.3957  0.2049 -0.1623  0.1740 -0.2140 -0.5387  0.2494  0.5735  0.0365
#>   0.2171  0.1484 -0.3474 -0.2530 -0.0537 -0.0158  0.1634  0.2539  0.1599
#>   0.2188  0.5042 -0.2527 -0.2904 -0.0624 -0.2509  0.4209  0.3128  0.1159
#> 
#> Columns 19 to 20  0.3670  0.2840
#>  -0.3208  0.0887
#>   0.3688  0.1098
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>