RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9329 -0.0240 -0.8482 -0.3821  0.1754  0.9083 -0.3376  0.6257 -0.8549
#>  -0.6309 -0.6268 -0.1387 -0.2014  0.6305  0.9157 -0.4437  0.2533  0.2765
#>  -0.1479 -0.3148  0.4226  0.1609  0.3460  0.2712  0.5418  0.2096 -0.1190
#> 
#> Columns 10 to 18 -0.3986 -0.1053 -0.4391 -0.1396 -0.1359 -0.0514  0.6393 -0.5692  0.2841
#>   0.1290  0.9025  0.7720  0.5989  0.5801  0.2927 -0.2252  0.5157  0.1109
#>  -0.9256  0.2502 -0.1818  0.1739 -0.0773 -0.3965 -0.1856 -0.4987 -0.3217
#> 
#> Columns 19 to 20  0.8484  0.2853
#>   0.3996  0.1043
#>   0.6285 -0.1987
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2402  0.5368  0.3122  0.2172 -0.2580  0.4095  0.0308 -0.6225 -0.7059
#>  -0.4098 -0.4457  0.2501  0.3120  0.1932 -0.0506 -0.4716 -0.1439 -0.1701
#>  -0.2571  0.1501 -0.3761 -0.6024  0.2789 -0.1247  0.1697  0.1998 -0.4166
#> 
#> Columns 10 to 18 -0.5976  0.3117 -0.0089  0.3367  0.7276  0.5570 -0.7113 -0.0230  0.1457
#>  -0.6019 -0.1922  0.3367  0.5660  0.3177 -0.6621  0.4596 -0.6628 -0.0370
#>  -0.2699 -0.2412 -0.5130 -0.3066 -0.0672  0.2629 -0.1835 -0.5900 -0.0707
#> 
#> Columns 19 to 20  0.0838  0.4717
#>   0.3038  0.5537
#>   0.4606  0.0836
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0372 -0.0324 -0.3871 -0.2867 -0.1745 -0.5482 -0.1199  0.4111 -0.2645
#>   0.1924  0.4658 -0.0368 -0.4334  0.0210  0.1747 -0.0490 -0.7259 -0.5629
#>  -0.1769  0.7473  0.2521 -0.2284  0.1290  0.6504  0.3020 -0.7078 -0.6752
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2875 -0.2859 -0.1564 -0.1556 -0.7671 -0.7366  0.1313  0.1220 -0.0048
#>  -0.2234  0.1492 -0.2114  0.5153 -0.1738 -0.4401 -0.6374 -0.0551  0.9363
#>  -0.5304 -0.7136  0.2009 -0.4580  0.0979 -0.1544  0.3792 -0.7208 -0.6587
#> 
#> Columns 10 to 18  0.2670 -0.2926  0.7458  0.7399 -0.7676 -0.2017  0.3594  0.2058  0.1317
#>   0.6275  0.4341 -0.4763  0.4522 -0.1153  0.0489  0.0325 -0.4893 -0.6393
#>  -0.0714 -0.2427  0.2785  0.2021 -0.4137 -0.2160  0.6111  0.1881  0.2070
#> 
#> Columns 19 to 20 -0.4694  0.1278
#>  -0.1052  0.1443
#>  -0.5294 -0.4104
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5375  0.4884 -0.3521 -0.6255  0.0294 -0.2246 -0.0855 -0.6660 -0.7661
#>  -0.0593  0.5610 -0.0546  0.0275 -0.0076  0.1508 -0.0337 -0.4561 -0.7298
#>   0.2084  0.1978  0.2554 -0.3575 -0.0520 -0.4125  0.4238 -0.7151 -0.4678
#> 
#> Columns 10 to 18 -0.0410  0.2158 -0.1721 -0.3340 -0.1317 -0.0238 -0.5671 -0.1621  0.1471
#>  -0.3656  0.1114  0.1104  0.3507  0.3422  0.3125 -0.3381 -0.2909 -0.0648
#>   0.2748  0.0453  0.4217 -0.0460  0.1703 -0.1320 -0.0142 -0.4382  0.3657
#> 
#> Columns 19 to 20 -0.1869  0.2796
#>   0.2078  0.0799
#>  -0.2632  0.3060
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>