RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3953 -0.6521 -0.2910  0.2256 -0.4320  0.5570 -0.2006 -0.0971 -0.2516
#>   0.2416 -0.5028 -0.2813 -0.4292 -0.4856  0.6315 -0.4857 -0.5439 -0.3336
#>  -0.9338 -0.1936 -0.4749 -0.4589 -0.5691  0.4015 -0.2195 -0.3767 -0.3026
#> 
#> Columns 10 to 18  0.7249  0.5959  0.6298 -0.3452  0.4943 -0.7125  0.6652 -0.7148  0.6404
#>  -0.3015 -0.7976  0.5217  0.7587 -0.7641  0.7445 -0.6530 -0.1582  0.5208
#>   0.8998  0.3854 -0.4008 -0.7016  0.5470  0.3702  0.2522  0.0209  0.0198
#> 
#> Columns 19 to 20  0.7686  0.2308
#>  -0.6439  0.2403
#>  -0.4366 -0.8801
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1279  0.0704 -0.5766 -0.0865 -0.3555  0.4155 -0.5973 -0.2542 -0.4473
#>  -0.5559  0.6754 -0.5301 -0.4520 -0.4839 -0.1216  0.4069 -0.5066  0.2686
#>   0.4040 -0.4767 -0.6956 -0.2213 -0.0650  0.5337 -0.3877 -0.1186 -0.3186
#> 
#> Columns 10 to 18  0.2088 -0.6487  0.0631  0.4857 -0.5790  0.3832  0.6065 -0.2992  0.6793
#>   0.5087 -0.2126 -0.5056 -0.2945  0.3867 -0.1674  0.6577 -0.1985  0.2251
#>  -0.2313 -0.1787  0.3478  0.5350  0.2215  0.4502  0.2204  0.0579  0.3170
#> 
#> Columns 19 to 20  0.7293  0.3220
#>  -0.5015 -0.0175
#>   0.2055  0.1973
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0964  0.2007 -0.5227 -0.0379 -0.3189  0.1211 -0.1538 -0.7579  0.0753
#>   0.3933 -0.7176 -0.7333 -0.2996  0.3309  0.4690  0.3881 -0.0310 -0.4266
#>  -0.3539  0.6686 -0.5512 -0.1623 -0.6027  0.2165 -0.1953 -0.4935 -0.0607
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3024  0.2908  0.2065 -0.0932  0.3328  0.3730  0.0531 -0.2020  0.3718
#>   0.1214 -0.5585 -0.2755 -0.3637 -0.3414 -0.0878 -0.5372 -0.0790 -0.0619
#>  -0.6752  0.5844 -0.7709  0.4249 -0.6923 -0.0063  0.2521 -0.1621  0.1264
#> 
#> Columns 10 to 18  0.3136  0.1081  0.1395 -0.7247 -0.0769 -0.8382  0.1869 -0.1829  0.0903
#>   0.1138  0.7020 -0.4070 -0.5966 -0.0534 -0.5595  0.0032  0.2724 -0.6804
#>   0.6079 -0.4044  0.2690  0.6542 -0.0467 -0.7953  0.4832 -0.6122 -0.7789
#> 
#> Columns 19 to 20 -0.3615  0.6369
#>   0.2772  0.2844
#>   0.3787 -0.2349
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2486 -0.5383 -0.6665  0.1760 -0.4703  0.5521 -0.2358 -0.5963 -0.0619
#>  -0.3375  0.2664 -0.5620 -0.0621 -0.4708  0.2650 -0.1789 -0.3003  0.0977
#>   0.0935 -0.2009 -0.5182 -0.0333  0.2556  0.4403  0.3143 -0.4251  0.0503
#> 
#> Columns 10 to 18 -0.0573 -0.3175 -0.2145  0.1737  0.1512  0.0789  0.2990  0.2135  0.3737
#>   0.2793 -0.3115 -0.5040  0.1313 -0.3142 -0.0809  0.6916  0.2547  0.3426
#>  -0.0359 -0.1506 -0.3453  0.4481 -0.2920  0.2327  0.5775  0.3140  0.5544
#> 
#> Columns 19 to 20  0.0847  0.5367
#>   0.0412  0.2638
#>   0.6569  0.2171
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>