RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1902 -0.4259  0.1382 -0.7295  0.4489 -0.4422  0.4150  0.4264  0.0194
#>  -0.1026  0.2931 -0.3252 -0.5708  0.3403  0.3802 -0.3328  0.6071 -0.3766
#>   0.1669  0.6574  0.2309 -0.7123  0.1358 -0.0172 -0.1288 -0.2023  0.5294
#> 
#> Columns 10 to 18 -0.5247 -0.2069 -0.1350  0.2854 -0.0514  0.8502  0.1614 -0.0987 -0.8971
#>   0.5940 -0.3556 -0.1841  0.6136  0.1539  0.2223 -0.1800 -0.1554 -0.0687
#>  -0.6976  0.7107 -0.8213 -0.2512 -0.1946 -0.0321  0.3463 -0.4951  0.4458
#> 
#> Columns 19 to 20 -0.5353  0.4535
#>   0.4583  0.2083
#>   0.2911 -0.7696
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4542 -0.4831 -0.1985 -0.6317 -0.1859 -0.1966  0.2921 -0.5556  0.1675
#>   0.4529  0.1871 -0.6532  0.1455 -0.3971  0.5989  0.2160 -0.1130  0.4262
#>   0.3771 -0.3651 -0.5852 -0.2887  0.5307  0.1562 -0.0192  0.3258  0.3185
#> 
#> Columns 10 to 18 -0.3052  0.4125 -0.0317 -0.0991  0.3287 -0.0129  0.1541 -0.6527  0.4538
#>  -0.2806  0.3686 -0.5027  0.0047  0.3200 -0.0630 -0.5416 -0.6321 -0.1383
#>  -0.6308 -0.0444  0.3309  0.2747  0.0389  0.2628  0.5980 -0.2720 -0.5520
#> 
#> Columns 19 to 20  0.0781 -0.6948
#>  -0.1323  0.1174
#>   0.4256 -0.2990
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0496 -0.4549 -0.1093 -0.1523  0.4143  0.1059 -0.2317 -0.1164 -0.1583
#>  -0.2525 -0.0479 -0.4806 -0.4728  0.4927  0.4739  0.0491  0.3954  0.4172
#>   0.0195 -0.3030 -0.1308 -0.3919  0.3326 -0.2617  0.3640 -0.4246  0.1041
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2764 -0.0020  0.3106 -0.3839 -0.1926  0.0977 -0.4381 -0.1180  0.6684
#>  -0.1540  0.5990 -0.5349  0.4721  0.3390  0.4473  0.3518  0.1374  0.4358
#>  -0.2024 -0.6218 -0.2678  0.4132 -0.2066  0.3018  0.0442  0.3291 -0.1370
#> 
#> Columns 10 to 18 -0.1734  0.7624 -0.1851 -0.2253  0.7564 -0.0224 -0.2000 -0.3595 -0.3277
#>   0.4735  0.0679  0.4389 -0.3540 -0.3649  0.0486  0.0768 -0.0188 -0.2198
#>   0.6889  0.3259  0.1591 -0.0347  0.5818  0.3403 -0.2665 -0.3020  0.0676
#> 
#> Columns 19 to 20  0.0081 -0.1098
#>  -0.2201  0.7042
#>  -0.2251  0.3995
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2061 -0.1738 -0.5691 -0.3259  0.2204  0.2356  0.1991  0.0694  0.4953
#>   0.2412 -0.6351 -0.3285 -0.6497  0.5215  0.4617 -0.2449  0.0266  0.2758
#>   0.2248 -0.6547 -0.5925 -0.5088  0.6337  0.4897  0.2217 -0.4456  0.1267
#> 
#> Columns 10 to 18 -0.3426  0.3709  0.0798  0.3808 -0.1111  0.5027  0.3809 -0.4349 -0.0752
#>  -0.1222  0.4530  0.0462  0.6031 -0.3440  0.1149 -0.3081 -0.6416  0.0519
#>  -0.5608  0.2917  0.2614  0.1188  0.2607  0.4760 -0.0280 -0.5954  0.1536
#> 
#> Columns 19 to 20  0.4987 -0.3996
#>   0.1394 -0.6173
#>   0.4376 -0.0268
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>