RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1192  0.3386  0.5778 -0.7339 -0.2945  0.4701  0.3932 -0.6342  0.3876
#>  -0.1887  0.3190 -0.4491  0.3147 -0.1468  0.0852  0.7137  0.8221 -0.0041
#>  -0.0282 -0.0635 -0.4160  0.2473  0.6733 -0.1137 -0.7572 -0.0011  0.8763
#> 
#> Columns 10 to 18 -0.1286 -0.1045  0.4120  0.6919  0.0888 -0.4873  0.1992  0.2951  0.1983
#>   0.5339 -0.4062 -0.3106 -0.1126 -0.3809 -0.6853  0.0933  0.7412  0.0884
#>   0.5914 -0.3550 -0.4385  0.4578 -0.2696 -0.5673 -0.1744  0.3565 -0.6226
#> 
#> Columns 19 to 20  0.4813  0.2733
#>  -0.7507 -0.2677
#>  -0.4375  0.5679
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0446 -0.5469  0.1215 -0.2480  0.3230  0.1967  0.4812 -0.2479  0.3346
#>  -0.1292  0.0926  0.3747 -0.0186 -0.1960  0.2895 -0.1595 -0.4079 -0.2467
#>  -0.3988 -0.2992  0.4960 -0.0100  0.3749 -0.2435  0.2703 -0.6647 -0.6291
#> 
#> Columns 10 to 18 -0.0137 -0.4618  0.4833  0.3746 -0.3498  0.1063  0.1293 -0.1740 -0.0812
#>  -0.3799  0.6120 -0.2704 -0.2368 -0.7134  0.2507 -0.2915 -0.2862 -0.2424
#>  -0.1394  0.3263  0.3947 -0.6887 -0.2941  0.3426 -0.3554 -0.2367  0.0407
#> 
#> Columns 19 to 20  0.0280  0.4394
#>   0.0139 -0.6768
#>   0.2338  0.0161
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3132 -0.2634 -0.1368  0.1075  0.5495 -0.3366  0.1285 -0.2303  0.4436
#>  -0.3232 -0.0825 -0.3932 -0.6565  0.5933 -0.5100 -0.4320 -0.6425 -0.0344
#>  -0.4596 -0.5267 -0.5909 -0.5665  0.6194 -0.1898  0.1943 -0.4661  0.3066
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2782 -0.1708 -0.0650 -0.2924  0.1751  0.1507 -0.0324 -0.0998 -0.2096
#>   0.1694  0.0201  0.5589  0.3211  0.3939  0.4478 -0.0641 -0.6514  0.2970
#>   0.0026  0.1915 -0.0028 -0.2640  0.4434 -0.2362  0.1399  0.5323  0.2545
#> 
#> Columns 10 to 18 -0.0697  0.4527 -0.6927  0.0487  0.2137 -0.3987  0.3086  0.0645  0.0132
#>   0.4899  0.0709 -0.3212  0.5404 -0.5043  0.2204  0.8727  0.9141 -0.6181
#>  -0.4757  0.1310  0.0530  0.4545  0.6969 -0.4728 -0.2178  0.2326 -0.1758
#> 
#> Columns 19 to 20  0.7734  0.6637
#>   0.5607  0.2594
#>   0.5110  0.0308
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0601 -0.1510 -0.0333 -0.0462  0.2189  0.2303  0.1753 -0.4166  0.0695
#>   0.0487 -0.1482 -0.2472 -0.0657  0.0052  0.3340  0.2309 -0.6729 -0.4158
#>  -0.1144 -0.1404  0.0266 -0.5321  0.2501  0.1285 -0.3528 -0.6767 -0.3309
#> 
#> Columns 10 to 18  0.1039 -0.1620  0.4885 -0.1440 -0.2594  0.0008  0.0501 -0.0324 -0.3209
#>   0.7600 -0.4844  0.1331 -0.3139  0.1457 -0.1991 -0.1686 -0.3754 -0.5955
#>  -0.0299 -0.0763  0.5105  0.1810 -0.4588  0.0543 -0.2371 -0.4251 -0.7305
#> 
#> Columns 19 to 20 -0.1089  0.2912
#>   0.2264  0.2613
#>  -0.0462  0.2119
#> [ CPUFloatType{2,3,20} ]
#>