RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3258 -0.0331  0.2495 -0.7825  0.1091  0.0854  0.4968 -0.5781 -0.8442
#>   0.3447  0.2804  0.7315 -0.6429  0.5032  0.5489 -0.2926 -0.4816  0.2310
#>  -0.2052 -0.5279 -0.9135 -0.2157 -0.4011  0.6599 -0.0141 -0.8793 -0.0600
#> 
#> Columns 10 to 18 -0.0576 -0.5038  0.3846 -0.6860  0.1134  0.8324  0.4495  0.2675  0.1513
#>   0.0976 -0.6778 -0.3186 -0.2573  0.4662  0.4096 -0.7030  0.8374  0.3875
#>  -0.2011  0.3958  0.2973 -0.6452  0.3467  0.1685 -0.3440  0.5566 -0.7802
#> 
#> Columns 19 to 20  0.4449 -0.4829
#>   0.8658  0.4492
#>  -0.8475  0.1248
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0141 -0.4920 -0.0648  0.0608  0.2653 -0.1886  0.2700 -0.0635 -0.3095
#>  -0.3348 -0.1944 -0.8784  0.1940 -0.5349  0.0642  0.4869 -0.3473 -0.3047
#>  -0.5767 -0.6408 -0.0971 -0.3417 -0.5009  0.7429  0.0626 -0.0143 -0.2139
#> 
#> Columns 10 to 18 -0.2079  0.2898 -0.3276  0.2292 -0.0824  0.0620  0.0017 -0.2763  0.1835
#>  -0.0799  0.3981 -0.4090  0.2451  0.0151 -0.0221  0.1424 -0.2792 -0.5219
#>   0.0059 -0.3055 -0.0428  0.2620  0.3467  0.6945 -0.5833 -0.0187 -0.1844
#> 
#> Columns 19 to 20  0.7479 -0.0315
#>   0.2185 -0.0856
#>  -0.3516  0.0357
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3262 -0.1182 -0.4404  0.0761 -0.2856  0.1399  0.3755 -0.3613 -0.4462
#>  -0.4616 -0.4630  0.3631 -0.0087  0.1391  0.1615 -0.0079 -0.0054 -0.0346
#>  -0.6280 -0.3308 -0.3355 -0.3748 -0.5715  0.1491  0.2545 -0.0946  0.0787
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7691  0.2700  0.5470 -0.4546  0.5829 -0.5337 -0.5547  0.5471  0.6877
#>   0.1224 -0.4313 -0.2110 -0.6302  0.3081 -0.5468  0.3990  0.2715  0.0038
#>  -0.2591 -0.3179 -0.1447 -0.2476  0.0171  0.1670 -0.7250  0.2912  0.3450
#> 
#> Columns 10 to 18 -0.1109  0.3836  0.3751 -0.5082 -0.1356 -0.4220  0.1957  0.1595  0.7783
#>  -0.4121  0.3476 -0.2011  0.2693 -0.0805 -0.6481  0.0157  0.1824  0.3664
#>   0.4743  0.2167 -0.5584 -0.8278 -0.2334 -0.5708 -0.6972  0.2629  0.6580
#> 
#> Columns 19 to 20  0.3715 -0.4167
#>  -0.3482 -0.1182
#>  -0.6703  0.2651
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4808  0.2179 -0.1993 -0.4783 -0.1749  0.1701  0.3667 -0.6266 -0.0548
#>  -0.3336 -0.5413  0.0900 -0.2091 -0.0231  0.0848  0.4298 -0.2053 -0.0062
#>   0.1085 -0.4125  0.2742 -0.5552 -0.0462 -0.5976  0.5108 -0.4667 -0.2743
#> 
#> Columns 10 to 18 -0.4876 -0.0320 -0.2605  0.0640  0.4170  0.3431 -0.4500  0.0596 -0.7270
#>  -0.3184  0.0106  0.0505 -0.1373  0.4873  0.6482  0.0155  0.0615 -0.3591
#>  -0.5530 -0.2184 -0.2210  0.1277 -0.3135  0.5866 -0.0942 -0.3558  0.1327
#> 
#> Columns 19 to 20  0.2653  0.2962
#>   0.2144  0.0577
#>   0.5104  0.2296
#> [ CPUFloatType{2,3,20} ]
#>