RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1436 -0.5247  0.6988 -0.4683 -0.0238 -0.6496  0.0873  0.3219  0.7388
#>   0.0648 -0.7765  0.2210 -0.3313 -0.7810 -0.7672 -0.7255 -0.8640 -0.3342
#>   0.9026 -0.9305 -0.8371 -0.8357 -0.8468 -0.0692  0.2329  0.5695 -0.7063
#> 
#> Columns 10 to 18 -0.2901  0.2014  0.9677 -0.1119 -0.3591  0.5117 -0.1035 -0.4929  0.9477
#>  -0.3528 -0.1483 -0.2471 -0.7358  0.4424  0.0674 -0.4293 -0.4214 -0.7492
#>  -0.6969 -0.7904 -0.0560 -0.2330  0.2391 -0.8500 -0.6528  0.4211  0.4182
#> 
#> Columns 19 to 20 -0.7425  0.6969
#>  -0.1310 -0.5784
#>   0.7423 -0.2744
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0533 -0.3423  0.6873 -0.4573 -0.0165  0.2927  0.0069 -0.3605  0.4590
#>   0.3218  0.5511 -0.4002 -0.2515 -0.2710 -0.1345 -0.3491  0.4922  0.4824
#>  -0.6637 -0.5568  0.0292 -0.5329  0.3394 -0.3456 -0.2455 -0.7323  0.2267
#> 
#> Columns 10 to 18 -0.2823 -0.2778  0.6163  0.5032  0.5900  0.0860 -0.3193 -0.7186  0.0638
#>   0.7853  0.0234  0.3324 -0.0938  0.6823 -0.7015  0.4444  0.4605  0.0336
#>  -0.2770 -0.3785  0.6860 -0.4310  0.6335  0.4002  0.1804 -0.4963  0.3988
#> 
#> Columns 19 to 20  0.5885  0.0289
#>   0.3423 -0.4725
#>  -0.7558 -0.2936
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4705 -0.1913  0.0608 -0.0589 -0.3725  0.0257  0.2811 -0.3297  0.5401
#>  -0.2148 -0.0994  0.1640  0.0945 -0.6404 -0.4388 -0.4290 -0.6864  0.0965
#>   0.4104  0.1994 -0.1682  0.1131 -0.7537  0.2075  0.2007  0.0683  0.6686
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1325  0.6219 -0.4092 -0.0816 -0.1200 -0.0949  0.5034 -0.0964  0.9270
#>   0.0699  0.5015  0.4635  0.0325  0.1575 -0.0676 -0.0739  0.4847  0.5910
#>  -0.5502 -0.2522 -0.0763  0.4251 -0.0980 -0.6598  0.4924  0.5375 -0.0310
#> 
#> Columns 10 to 18  0.3628 -0.1501  0.5448  0.4775 -0.4749  0.8078 -0.6386 -0.5763 -0.6614
#>   0.5962  0.4069 -0.4687 -0.0020  0.1156  0.1155  0.2741 -0.2981  0.0826
#>  -0.4149 -0.1039  0.4462  0.0830  0.2042  0.4048  0.8724  0.3996  0.1871
#> 
#> Columns 19 to 20 -0.2650  0.1300
#>  -0.2002 -0.4323
#>  -0.2783  0.3378
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1701  0.0866 -0.1414  0.0655 -0.4262 -0.2951  0.1301 -0.5969  0.5628
#>   0.1589  0.0534 -0.0715  0.1700  0.1394 -0.0252 -0.1966 -0.5224  0.4640
#>   0.1029 -0.1752  0.1471 -0.0234 -0.6939 -0.3672 -0.1336 -0.2248  0.5523
#> 
#> Columns 10 to 18 -0.0537 -0.1369  0.4948  0.1082  0.2072 -0.1386 -0.4155  0.0376 -0.1377
#>   0.4145  0.0102  0.5452 -0.4040  0.2365  0.4903  0.0955 -0.0533  0.1467
#>   0.2709 -0.2458  0.2570  0.0102  0.3620 -0.1759  0.1151 -0.4870  0.2677
#> 
#> Columns 19 to 20  0.3544 -0.0503
#>  -0.0776 -0.4562
#>  -0.3353 -0.1150
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>