RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1472 -0.0970 -0.3496 -0.8381 -0.3380  0.4668 -0.4055  0.4050 -0.8870
#>  -0.5332 -0.4966 -0.1355 -0.1370 -0.0969 -0.8234 -0.0885  0.2070  0.3812
#>  -0.6407  0.4761  0.2915 -0.7853  0.1055 -0.7407 -0.2014  0.2888  0.4075
#> 
#> Columns 10 to 18 -0.9358 -0.7628  0.7143  0.1121  0.6631  0.9886 -0.7411 -0.3596 -0.4021
#>  -0.6603 -0.2390 -0.4003  0.4240 -0.1189  0.8659 -0.3029 -0.3222  0.6798
#>  -0.8152  0.4912  0.0979  0.5257  0.2083 -0.8175  0.1199 -0.1093 -0.0079
#> 
#> Columns 19 to 20  0.5610 -0.8259
#>   0.0620  0.0573
#>   0.0853 -0.6521
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0271 -0.3470 -0.5212 -0.3940  0.2895  0.3320  0.2023  0.0092  0.4081
#>   0.6473  0.0986 -0.5528 -0.3799  0.3566  0.0746 -0.3437  0.2609 -0.1522
#>   0.0421 -0.2967 -0.0190 -0.4469  0.2568  0.2339  0.5395  0.2658  0.4897
#> 
#> Columns 10 to 18  0.1436 -0.7240 -0.3410 -0.7373 -0.0115  0.0103 -0.6570  0.5193  0.0187
#>  -0.3782 -0.4446  0.0814 -0.0951  0.0382  0.2447 -0.7720 -0.1294 -0.2111
#>   0.5757 -0.2954  0.3787  0.1489  0.0446 -0.3623 -0.5161  0.5413 -0.6218
#> 
#> Columns 19 to 20  0.8241 -0.3723
#>   0.4808 -0.5097
#>  -0.2135 -0.3995
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0333 -0.3454  0.0047 -0.4091 -0.3708 -0.1738  0.0559 -0.1141 -0.6677
#>   0.0799 -0.6875 -0.4219  0.0171  0.3052  0.4339 -0.0923  0.2222 -0.2735
#>  -0.5179 -0.3487  0.2467  0.2039 -0.3499  0.1246  0.2236 -0.2543  0.6846
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0369  0.2869 -0.7571 -0.1840 -0.5368  0.2961 -0.5835 -0.3892  0.1660
#>   0.2071 -0.6692 -0.7282 -0.0758  0.5972 -0.2338  0.2121 -0.3458  0.4003
#>   0.0956  0.5070 -0.5499  0.1278 -0.4782  0.1026 -0.1916 -0.8370 -0.5715
#> 
#> Columns 10 to 18 -0.2973 -0.3454  0.2969  0.2197  0.5893  0.2301 -0.4170  0.1228 -0.2001
#>  -0.0528 -0.0516 -0.0311 -0.0404 -0.5803  0.1425 -0.0830 -0.6796 -0.7060
#>  -0.2763  0.5762  0.1646 -0.1948  0.4185  0.7118  0.2742 -0.6863 -0.3617
#> 
#> Columns 19 to 20 -0.1105 -0.0917
#>   0.2958 -0.0533
#>   0.1432 -0.2829
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1149 -0.0902  0.4979 -0.6357  0.2267 -0.1996  0.0218  0.3882  0.1437
#>  -0.0375 -0.1308 -0.4402 -0.4620 -0.1540  0.1807 -0.0746  0.1255  0.2380
#>  -0.4145 -0.4729  0.2998 -0.3064 -0.2698  0.0741  0.4181 -0.1791  0.5174
#> 
#> Columns 10 to 18  0.0764 -0.4868  0.0828  0.0926  0.1832  0.2321 -0.4256 -0.4154 -0.0382
#>   0.0757  0.3976  0.1018 -0.1953 -0.1665  0.0645 -0.1737 -0.2613 -0.0206
#>  -0.2989 -0.0048 -0.0208  0.4061 -0.2573  0.0314 -0.4099 -0.2569  0.1312
#> 
#> Columns 19 to 20  0.2697 -0.3390
#>  -0.0217  0.1570
#>  -0.5177  0.0164
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>