RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2296 -0.2165  0.1527 -0.6248 -0.9343  0.6963  0.4189  0.6768 -0.1335
#>  -0.8057  0.9446 -0.6518  0.5591 -0.0486 -0.4770  0.2831  0.1846  0.7485
#>  -0.5684 -0.7321  0.1425 -0.3827 -0.3968 -0.5226  0.1496  0.6009 -0.7993
#> 
#> Columns 10 to 18 -0.4661 -0.6567 -0.2890  0.7469  0.7330 -0.3519 -0.5628  0.8097 -0.4385
#>   0.0194 -0.7904  0.9427 -0.3930  0.2815  0.8304 -0.3246  0.8383  0.0645
#>   0.6382  0.3941 -0.4935  0.3031 -0.4884  0.1128  0.5901  0.6731  0.1187
#> 
#> Columns 19 to 20 -0.5405 -0.2334
#>   0.5901 -0.3574
#>  -0.3316 -0.4676
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3128  0.1522 -0.3244 -0.8348 -0.0924  0.0384  0.4868 -0.4201  0.4081
#>   0.0204  0.6117 -0.3489 -0.4115 -0.3296  0.3299  0.3743  0.2617  0.2002
#>  -0.2701  0.0282 -0.6428 -0.5488  0.1144 -0.0531 -0.3821 -0.0486  0.2438
#> 
#> Columns 10 to 18  0.6324 -0.3814 -0.3493  0.2596 -0.0971 -0.4915  0.1611  0.2039  0.5529
#>   0.4239  0.0162  0.4143  0.2886  0.2991 -0.5396 -0.6201  0.6891  0.0865
#>   0.3123 -0.1653 -0.1732  0.0577  0.1772 -0.1695  0.3680  0.6045  0.0257
#> 
#> Columns 19 to 20  0.7781 -0.1798
#>  -0.4150 -0.0342
#>   0.3011 -0.4057
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4807  0.2098 -0.5957  0.2669  0.1152 -0.1355  0.1249  0.0740 -0.0414
#>  -0.3599  0.6655 -0.6865 -0.2792 -0.3046  0.1168  0.5995  0.1230  0.2003
#>  -0.5364  0.3946 -0.4533 -0.2514 -0.3153  0.1129  0.2216 -0.2090  0.4557
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2667  0.7049  0.2158 -0.1848  0.2632  0.0134 -0.2312  0.3646 -0.3817
#>   0.4315 -0.2927 -0.1809 -0.1944  0.3666 -0.1744 -0.5905  0.3274  0.6545
#>   0.0502  0.3470  0.6771 -0.7468  0.5672 -0.8077  0.4228  0.8647  0.2465
#> 
#> Columns 10 to 18  0.5099 -0.7119  0.4247 -0.1557 -0.0472  0.0589  0.7025  0.5159  0.1425
#>  -0.1574  0.2416  0.3858 -0.6948 -0.5558  0.1878 -0.5273 -0.1684  0.4746
#>   0.5754  0.4768  0.3590 -0.6953  0.2753 -0.3086  0.5404  0.2047 -0.6690
#> 
#> Columns 19 to 20  0.6863  0.4834
#>   0.6934  0.5110
#>   0.7641  0.1408
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0257  0.6942 -0.5338 -0.6709 -0.4548  0.3050  0.6509 -0.0392 -0.0900
#>  -0.4772  0.6912 -0.6423 -0.3219 -0.4922 -0.0941  0.6516  0.3925 -0.1000
#>  -0.3908  0.6562 -0.6346 -0.1294 -0.4073  0.5028  0.6710  0.2987  0.5049
#> 
#> Columns 10 to 18  0.4996 -0.1613  0.2518  0.1535 -0.0782 -0.4393 -0.2183  0.2089  0.1589
#>   0.3849 -0.4002  0.5302  0.4701  0.3199  0.0059 -0.3561  0.4677  0.0772
#>   0.6642 -0.3892  0.5987 -0.0989 -0.2014 -0.2104 -0.7206  0.3787  0.2139
#> 
#> Columns 19 to 20 -0.0243  0.0131
#>   0.3648  0.1590
#>   0.3832  0.0262
#> [ CPUFloatType{2,3,20} ]
#>