RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4436 -0.6173 -0.2167  0.8681 -0.1721  0.4085 -0.2902 -0.5553 -0.0399
#>   0.6031  0.1699  0.1545 -0.7247 -0.2512  0.4804 -0.4506  0.3209 -0.3383
#>  -0.1659 -0.4261  0.3871  0.4948  0.1319 -0.7596  0.2993 -0.9051  0.4003
#> 
#> Columns 10 to 18 -0.1258 -0.1436  0.7969 -0.0700  0.0001  0.4612 -0.6119  0.0411 -0.6822
#>  -0.4168 -0.3439 -0.0873  0.0234 -0.7198  0.2023  0.3208  0.2953 -0.2954
#>  -0.0040 -0.1527 -0.7109 -0.0945  0.2268  0.0296 -0.8632 -0.5509 -0.3242
#> 
#> Columns 19 to 20 -0.3946 -0.1427
#>   0.1965 -0.7774
#>   0.9628  0.4399
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6670  0.3167  0.0391 -0.2603  0.4079  0.2168 -0.3807 -0.3766 -0.2224
#>  -0.5243  0.0418 -0.5513  0.7368  0.2897  0.2563 -0.1577 -0.3406  0.1752
#>   0.0997 -0.1497  0.1457  0.2219  0.4828  0.2781  0.0625 -0.3478 -0.2528
#> 
#> Columns 10 to 18 -0.5327 -0.1261  0.0364  0.6054  0.0324  0.4629 -0.0157  0.1578  0.1426
#>  -0.2462 -0.0889  0.0482  0.0034 -0.1921  0.1937  0.1576  0.0041 -0.4576
#>   0.2183  0.7445  0.0737  0.0163 -0.3053  0.2845  0.1326 -0.2773 -0.1975
#> 
#> Columns 19 to 20 -0.3219  0.1726
#>  -0.5648  0.0798
#>   0.2497 -0.0384
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1129  0.6373 -0.0016  0.4242  0.2464  0.4114 -0.5300 -0.3887  0.1906
#>   0.4227  0.6096 -0.0379 -0.1901 -0.2977 -0.0558 -0.4949 -0.3824  0.0510
#>   0.1108 -0.2468 -0.1444 -0.2662 -0.0962  0.0307 -0.1253 -0.0241 -0.1666
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5167  0.0111  0.0753  0.2059 -0.5648  0.5450 -0.2922 -0.3481 -0.2144
#>  -0.5683 -0.6987  0.3572  0.3050 -0.5914  0.6016 -0.3118 -0.4608 -0.2045
#>   0.3201  0.1749 -0.1057  0.4021  0.6796 -0.3507  0.1766  0.1744  0.7046
#> 
#> Columns 10 to 18 -0.5276 -0.6425  0.2109  0.0145  0.1154 -0.3994  0.6904  0.0371  0.3507
#>  -0.5785 -0.2384  0.6422 -0.5794  0.0844 -0.2348  0.5493  0.0340  0.2219
#>  -0.3156 -0.2288  0.5261 -0.4275 -0.5845  0.1910 -0.5795  0.3838 -0.5316
#> 
#> Columns 19 to 20 -0.2221 -0.5326
#>  -0.8029 -0.5650
#>  -0.4094  0.6799
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7119  0.3073  0.4265  0.2489 -0.0588 -0.5760 -0.3004 -0.4348  0.2747
#>  -0.4955 -0.0152  0.5230  0.3614  0.3360 -0.7246 -0.4849 -0.4649  0.3183
#>   0.2768 -0.0078 -0.1177  0.2987  0.1183  0.1690 -0.2661 -0.2346  0.0589
#> 
#> Columns 10 to 18  0.3275  0.5355 -0.3257  0.3463 -0.1090 -0.0955  0.2295  0.1466 -0.1892
#>  -0.1190  0.5585 -0.3725  0.2598  0.1229  0.1299  0.3086  0.1493 -0.2258
#>  -0.4452  0.1408  0.1762  0.4630 -0.0467  0.4685 -0.4294  0.1654 -0.2000
#> 
#> Columns 19 to 20  0.4221 -0.3882
#>   0.4527 -0.7060
#>  -0.2352  0.2929
#> [ CPUFloatType{2,3,20} ]
#>