RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6558 -0.8875 -0.5224  0.4442 -0.4272  0.9335  0.4252  0.3032 -0.5899
#>  -0.1389  0.3966 -0.0599  0.5148 -0.6176  0.6398  0.3150 -0.3014 -0.7263
#>   0.2661 -0.0412  0.0780  0.7290  0.5015  0.4397  0.0032  0.3803 -0.3515
#> 
#> Columns 10 to 18  0.7607 -0.7161  0.9546  0.4141  0.3603 -0.2209 -0.3220  0.0519 -0.4539
#>  -0.0758 -0.2204 -0.2899  0.4955  0.8994  0.1349 -0.0359  0.3513 -0.1234
#>   0.2116 -0.7278 -0.6673 -0.2830 -0.4986 -0.7972  0.1699  0.7269  0.5157
#> 
#> Columns 19 to 20 -0.6204  0.9138
#>   0.8753  0.7105
#>   0.0982 -0.1532
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2674 -0.1709 -0.5547  0.5433 -0.2079 -0.0194 -0.3275  0.3214  0.2906
#>   0.5022 -0.1823 -0.6560 -0.0324 -0.1494  0.2960  0.0668  0.0343 -0.5908
#>   0.1833 -0.0048  0.5237 -0.4981 -0.3739 -0.7435 -0.5736  0.0095 -0.4174
#> 
#> Columns 10 to 18  0.0058 -0.2467  0.4811 -0.3249  0.5701 -0.2784 -0.3228  0.1046  0.5921
#>   0.0249 -0.5886  0.1275  0.3875  0.2364 -0.4301  0.4270  0.7827 -0.3282
#>  -0.2350 -0.4213  0.2189 -0.5306  0.0611  0.1182 -0.2269  0.1120 -0.1443
#> 
#> Columns 19 to 20  0.0323 -0.4184
#>   0.1207  0.0129
#>   0.0938 -0.3374
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3777 -0.5127 -0.2226  0.5420  0.2089  0.2445 -0.2216  0.3755 -0.3667
#>   0.1435 -0.6348 -0.4188  0.1456 -0.2935  0.4027  0.3048  0.2513 -0.2121
#>   0.3174  0.0811 -0.2041 -0.5476  0.2576  0.2948  0.3469  0.4904 -0.1713
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1758  0.3284  0.2435 -0.6769 -0.0019  0.2629 -0.1551  0.5045 -0.6259
#>  -0.4812  0.1379 -0.0776 -0.5527 -0.3092  0.1993  0.2829  0.0012 -0.0153
#>   0.4449  0.1272 -0.2470  0.2397 -0.1753  0.7446 -0.5201  0.6978  0.0258
#> 
#> Columns 10 to 18  0.3030  0.6395 -0.2630  0.3002 -0.3638  0.6492  0.2229 -0.3398  0.3265
#>   0.2575  0.4717 -0.7240  0.4272  0.4574 -0.4412 -0.1803 -0.5857  0.3556
#>   0.2300  0.7415 -0.6127 -0.4266  0.0730  0.5151 -0.3765  0.1683 -0.6087
#> 
#> Columns 19 to 20 -0.3179  0.0420
#>   0.6533 -0.4120
#>  -0.1912  0.2723
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2427 -0.4572 -0.0786 -0.2674 -0.0070  0.2800  0.2172  0.2409  0.2452
#>   0.4624 -0.5027 -0.0304 -0.4854 -0.2151 -0.4511 -0.1941 -0.0575 -0.1354
#>   0.0426 -0.4301 -0.1293  0.3991  0.2829  0.1888  0.2008  0.3919  0.0161
#> 
#> Columns 10 to 18 -0.0727 -0.3939  0.4205  0.1533  0.4878  0.1935  0.0556 -0.0365 -0.5836
#>  -0.2264 -0.1008  0.1874 -0.6757  0.1214 -0.0440 -0.0516  0.2717 -0.1311
#>  -0.1558 -0.5440  0.2887 -0.0580  0.2737 -0.2546  0.3301  0.1773  0.0690
#> 
#> Columns 19 to 20  0.0466  0.2400
#>   0.0851 -0.3979
#>   0.1271  0.0269
#> [ CPUFloatType{2,3,20} ]
#>