RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0763 -0.4442  0.0314 -0.7162  0.4794 -0.5000  0.4377  0.6974  0.4072
#>   0.4637 -0.5063 -0.4083 -0.2359  0.6784 -0.2432  0.1823  0.2819  0.2557
#>  -0.2061  0.7759 -0.4293 -0.7793  0.6311  0.3288  0.8294 -0.7188 -0.7372
#> 
#> Columns 10 to 18 -0.8537 -0.1339 -0.1447  0.9397 -0.2173  0.4676  0.7704 -0.4687 -0.9412
#>  -0.2948 -0.7191 -0.2823  0.1093 -0.6899 -0.5638  0.7759 -0.8089 -0.8274
#>  -0.0883 -0.1925 -0.4620  0.4403  0.0472 -0.0210  0.6600 -0.4213  0.3333
#> 
#> Columns 19 to 20 -0.5245 -0.3928
#>   0.2587 -0.2539
#>  -0.5183  0.3598
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3280  0.0066  0.1135  0.2629 -0.1508  0.2119 -0.1384  0.8356  0.3934
#>   0.0802 -0.7309  0.3794  0.0541  0.5167 -0.0029  0.5787 -0.2465  0.0786
#>   0.0215  0.5669 -0.5192  0.4757 -0.0180  0.0602 -0.3048  0.7348 -0.2515
#> 
#> Columns 10 to 18 -0.2657  0.0005  0.3742 -0.2738  0.1888 -0.3675  0.6646  0.0525 -0.4804
#>  -0.7246 -0.2256 -0.4444  0.1929  0.0984  0.0592  0.5078  0.0461 -0.3311
#>  -0.1127  0.0122  0.0391  0.3470 -0.2475 -0.1850  0.4424 -0.6359 -0.1846
#> 
#> Columns 19 to 20  0.1446 -0.3574
#>  -0.0752 -0.3192
#>   0.7425  0.3973
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0359  0.2975  0.5368  0.3736 -0.0403  0.3347  0.4504  0.6290 -0.2847
#>  -0.0396  0.2086  0.4071  0.2502  0.0807  0.1553  0.2686  0.5889  0.1841
#>   0.1967  0.1881  0.3101  0.2192  0.3913  0.0135  0.6220 -0.4967 -0.5515
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1620 -0.1872  0.5139  0.3588 -0.4856  0.1164 -0.6429 -0.4577  0.5963
#>  -0.3957  0.6611 -0.8162  0.6115  0.5965  0.5495 -0.1724  0.0385 -0.3391
#>  -0.6269 -0.0774 -0.5079 -0.2435 -0.2142  0.6939 -0.5760 -0.7826  0.5511
#> 
#> Columns 10 to 18  0.5798 -0.0697 -0.4034 -0.0542  0.0524  0.1666  0.4722 -0.1561 -0.0584
#>  -0.0943 -0.2712 -0.3032 -0.4773 -0.0720 -0.5213 -0.6228  0.3386 -0.6646
#>   0.6576 -0.5628 -0.1896  0.2752 -0.2726 -0.0260 -0.3464 -0.4393 -0.4643
#> 
#> Columns 19 to 20  0.5968  0.2468
#>  -0.4079 -0.3241
#>  -0.6361  0.7021
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1666  0.0606  0.5407  0.4539  0.2421  0.4512  0.2133 -0.0487 -0.3761
#>   0.3003  0.3626  0.2212  0.5539 -0.2842  0.6016  0.2403  0.6706  0.0905
#>   0.4056 -0.4961  0.6891 -0.1210 -0.1781  0.2381  0.5620  0.4945 -0.5315
#> 
#> Columns 10 to 18 -0.5909 -0.3466 -0.3395 -0.2090  0.0817  0.1214 -0.2429 -0.1491  0.2526
#>  -0.5241 -0.6035 -0.2289  0.2018  0.4819 -0.3961 -0.0562 -0.2377 -0.2150
#>  -0.0843  0.0505 -0.6743 -0.1359  0.2928  0.2883  0.1153 -0.2411 -0.5036
#> 
#> Columns 19 to 20  0.6991  0.0396
#>   0.4223  0.0912
#>   0.2372 -0.1074
#> [ CPUFloatType{2,3,20} ]
#>