RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3690  0.1857 -0.8133  0.4152  0.0839  0.4827  0.0527  0.8591 -0.8048
#>  -0.8421 -0.1236 -0.3967  0.3669 -0.3461 -0.4389  0.1496  0.2327 -0.5966
#>   0.7542 -0.5141  0.7619  0.0806  0.8760  0.3906  0.7815 -0.3520 -0.5362
#> 
#> Columns 10 to 18 -0.9088 -0.3775 -0.6199  0.4903  0.4964  0.0726  0.3176  0.3586  0.2662
#>  -0.8676 -0.4074  0.5423  0.0860 -0.4344 -0.3115  0.2712 -0.7363  0.5386
#>   0.6232 -0.2167 -0.1889  0.8572 -0.0076  0.3263  0.5052 -0.0989  0.0058
#> 
#> Columns 19 to 20  0.7702  0.6941
#>  -0.2111 -0.2843
#>   0.1654  0.1040
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3073 -0.2821 -0.3473 -0.5953  0.3979  0.4960  0.1409 -0.2351 -0.1871
#>   0.1226 -0.5144 -0.1499 -0.2645  0.2303  0.3660  0.7626 -0.0466  0.1553
#>  -0.1479 -0.2775 -0.3587 -0.5378 -0.0342 -0.0401 -0.2343 -0.1617  0.4895
#> 
#> Columns 10 to 18 -0.7474 -0.5681  0.2650  0.2430 -0.5348  0.1845 -0.0296 -0.6056  0.1887
#>  -0.3105  0.2422 -0.3333  0.4755 -0.0969 -0.3007  0.4319  0.0471  0.6496
#>  -0.1586  0.1896  0.5643  0.7977 -0.0593 -0.1704  0.4512 -0.1255  0.5515
#> 
#> Columns 19 to 20 -0.0620  0.0759
#>   0.6420 -0.7662
#>  -0.1566 -0.0929
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2147 -0.0569 -0.2626  0.0230  0.0635  0.4218  0.1896 -0.3812  0.0016
#>  -0.6078  0.1476  0.0353 -0.1468 -0.4638 -0.0078  0.0215 -0.0073 -0.2408
#>   0.2476 -0.1027 -0.2683 -0.5061 -0.5290 -0.1285 -0.0406  0.0587 -0.4480
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4836  0.6104 -0.5605  0.5270 -0.1438  0.2817 -0.2778 -0.8173  0.4292
#>   0.5814  0.0088 -0.6950  0.4164  0.2760 -0.7650 -0.4728 -0.2806  0.1626
#>  -0.2642  0.4764 -0.4272 -0.5460 -0.6122  0.7617  0.4883 -0.7973  0.3555
#> 
#> Columns 10 to 18 -0.1930  0.4810 -0.0818  0.2929  0.3619  0.0527  0.1591 -0.0147 -0.0072
#>  -0.6661  0.2478 -0.1969 -0.0044 -0.2346 -0.2401  0.8508  0.6571 -0.2956
#>   0.5084  0.6266 -0.0002  0.1937  0.3030 -0.3104 -0.3588  0.1171  0.1654
#> 
#> Columns 19 to 20  0.3496 -0.7528
#>   0.0310  0.0369
#>   0.4864 -0.1773
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2860 -0.0064 -0.1313 -0.1903 -0.1495  0.1920  0.0267 -0.3078  0.0542
#>  -0.0706 -0.1869 -0.1360 -0.5697 -0.6042  0.2688 -0.2876 -0.2495  0.1517
#>  -0.1859 -0.1549  0.0756 -0.0508  0.2133 -0.0819  0.3817 -0.3710 -0.4933
#> 
#> Columns 10 to 18 -0.2186 -0.0049  0.0526  0.5281 -0.4257 -0.3517  0.4465  0.1252 -0.0412
#>  -0.2890 -0.1274  0.6679  0.4868  0.1667 -0.1850  0.5886  0.0537  0.4930
#>  -0.2297  0.4058 -0.4550  0.5647 -0.4599 -0.2248 -0.2086 -0.4161 -0.0082
#> 
#> Columns 19 to 20 -0.2189 -0.4822
#>  -0.1271 -0.1606
#>  -0.0371 -0.0640
#> [ CPUFloatType{2,3,20} ]
#>