RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0346  0.4706 -0.1155  0.7167  0.3321 -0.5457  0.1790  0.4632 -0.0870
#>   0.7175 -0.7561  0.2309 -0.8624  0.7036 -0.3179  0.6661  0.6094  0.3502
#>   0.3325  0.2831 -0.5509  0.4825  0.3294 -0.9607  0.6906 -0.3606 -0.5381
#> 
#> Columns 10 to 18 -0.8481  0.3662 -0.2667  0.1903 -0.3183  0.9315 -0.6692  0.7069  0.3291
#>   0.7084  0.3427 -0.0597  0.0457 -0.0119  0.3682 -0.4244  0.2426  0.5139
#>  -0.3592  0.4504  0.0715 -0.4019 -0.2070  0.9035 -0.4154  0.8592  0.7887
#> 
#> Columns 19 to 20 -0.6484  0.2201
#>   0.3874 -0.7758
#>  -0.4519 -0.0518
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2301  0.1493 -0.0708 -0.0641 -0.2646 -0.0820  0.5657 -0.5010 -0.3032
#>   0.1944 -0.0835 -0.4373  0.3635  0.1463 -0.2408  0.5347  0.5118 -0.5244
#>  -0.2187 -0.3349  0.0806 -0.4910 -0.0200 -0.0548  0.1262 -0.1075 -0.2865
#> 
#> Columns 10 to 18  0.3063  0.4190 -0.4900 -0.1047  0.3409  0.4394 -0.2347  0.6075 -0.0323
#>   0.1668  0.5443  0.2649 -0.6840  0.1068  0.1189 -0.6089  0.4187  0.2292
#>  -0.0893  0.7073 -0.0905  0.4664  0.0037  0.7765  0.1103  0.2764  0.0970
#> 
#> Columns 19 to 20 -0.5817  0.1578
#>  -0.5056  0.3633
#>  -0.6877 -0.1809
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0698 -0.2336 -0.4411  0.5594  0.2345 -0.1256  0.7823 -0.3039 -0.0369
#>   0.2252 -0.4383  0.0106 -0.2742 -0.1136  0.2841  0.3978  0.0799 -0.0212
#>   0.5171  0.4354 -0.5213  0.1454  0.3075 -0.5983  0.8642 -0.0591 -0.6311
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3563 -0.4003  0.0710  0.0695  0.7918  0.5837  0.7972  0.3727  0.6598
#>   0.4471  0.0794 -0.3482 -0.0956  0.4079  0.1690  0.3351 -0.3143 -0.3735
#>   0.4409 -0.4800  0.2285 -0.1612 -0.0061 -0.1875  0.1291 -0.2488  0.8258
#> 
#> Columns 10 to 18 -0.5938  0.6519  0.4013 -0.6472  0.8836  0.3333  0.6769  0.0369 -0.4376
#>  -0.7495 -0.6588 -0.5189 -0.1306 -0.3542  0.5365  0.1841  0.6383  0.2031
#>  -0.7407 -0.9371  0.0524  0.3972 -0.3343  0.6568  0.3464 -0.0062 -0.0856
#> 
#> Columns 19 to 20 -0.0018 -0.7192
#>   0.0246 -0.3200
#>   0.0259 -0.7779
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4483 -0.3750  0.1100  0.2498  0.7002 -0.0089  0.5518 -0.8006  0.2907
#>   0.2152  0.1054  0.0174  0.3914 -0.1880  0.0603  0.2874  0.2042 -0.2715
#>  -0.0730  0.2294 -0.1013  0.0343 -0.1637  0.0970  0.1980 -0.3311  0.1407
#> 
#> Columns 10 to 18  0.2316  0.3717  0.0092  0.5631 -0.2846  0.5784 -0.2273  0.1481  0.5043
#>   0.0646  0.2972 -0.4418 -0.4158 -0.4297  0.2126 -0.3619  0.6050  0.3824
#>  -0.1943  0.7218 -0.4986 -0.1164  0.1445  0.5844  0.1246  0.4364  0.1163
#> 
#> Columns 19 to 20  0.0131 -0.0318
#>  -0.4076  0.1948
#>  -0.6250  0.0441
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>