RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0168 -0.3302  0.2378  0.6402 -0.7878  0.1271  0.0388  0.9273 -0.8790
#>  -0.2201 -0.0166 -0.4212 -0.2101  0.0711  0.8186  0.0110  0.8199 -0.6088
#>  -0.0802 -0.2349 -0.0361  0.2801  0.6815 -0.6636 -0.0078  0.8501 -0.1928
#> 
#> Columns 10 to 18 -0.0431 -0.1697  0.0599  0.7064  0.2050  0.9359 -0.3032  0.4972  0.7136
#>   0.2442 -0.5033  0.3633  0.3144  0.6908  0.6259 -0.4471 -0.5273 -0.3185
#>  -0.4190  0.4711 -0.2177  0.7283  0.6718  0.8576 -0.7029  0.6138  0.6904
#> 
#> Columns 19 to 20  0.2885  0.8796
#>   0.6527  0.8169
#>   0.6622  0.5291
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1239 -0.4014  0.3168  0.5694 -0.2039 -0.4453  0.0230  0.2555 -0.2689
#>  -0.3921 -0.4603 -0.4434  0.2250  0.5678 -0.4403  0.4348  0.3022 -0.2934
#>  -0.2347  0.0353 -0.2954  0.0113 -0.3578 -0.6361 -0.1329  0.0229  0.2043
#> 
#> Columns 10 to 18 -0.0310  0.3167 -0.2544  0.5086  0.4548  0.1730 -0.1853  0.1966 -0.1375
#>  -0.0969 -0.0619 -0.0172 -0.1716  0.2054  0.1681 -0.3325  0.2138  0.3985
#>  -0.2318 -0.3226 -0.0406 -0.1327  0.6546  0.3837 -0.6697 -0.4030  0.3870
#> 
#> Columns 19 to 20  0.5614  0.2793
#>   0.4015 -0.5407
#>   0.6471  0.8439
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3322 -0.2613 -0.3892 -0.3839  0.4298  0.0542 -0.2147  0.5962  0.0872
#>  -0.2710  0.4511 -0.3577 -0.1897 -0.2063 -0.0994 -0.0526  0.1624  0.4489
#>  -0.4445 -0.0762  0.2136  0.0449  0.5251 -0.2559  0.4539  0.2783 -0.1895
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1505  0.3148 -0.5682 -0.4134  0.1622 -0.2883 -0.7267  0.1211 -0.3651
#>   0.4058  0.1832  0.4754  0.4288  0.1461  0.7311 -0.0987 -0.1957  0.0332
#>   0.1181  0.5669  0.0463  0.3540  0.5040  0.8202  0.3181  0.7677  0.6488
#> 
#> Columns 10 to 18 -0.1284  0.0291 -0.8428  0.5444  0.1576  0.5915 -0.3511 -0.8051 -0.5304
#>   0.1576  0.6395 -0.4542 -0.2301 -0.3320  0.6572  0.2024 -0.4896 -0.1961
#>   0.2736  0.5658  0.0024  0.2533 -0.3845  0.4735  0.1760 -0.3188  0.1039
#> 
#> Columns 19 to 20  0.8302 -0.3871
#>   0.7727  0.0671
#>  -0.7517 -0.7136
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0098 -0.1751  0.4475  0.3268  0.0368 -0.6519  0.1512  0.3688 -0.2272
#>  -0.4626 -0.2467  0.0042 -0.0764 -0.2358 -0.2235  0.0685  0.4947  0.0681
#>  -0.2709 -0.1643  0.1592 -0.2848  0.2139 -0.2351  0.0422  0.1662  0.1659
#> 
#> Columns 10 to 18 -0.3940  0.0651  0.0844  0.3548  0.4729  0.2970 -0.5379  0.0705  0.0900
#>   0.0283 -0.2299 -0.0780  0.5748  0.6782  0.3873 -0.1700 -0.1119  0.3940
#>   0.0221 -0.5085 -0.1979  0.1008  0.2544  0.6710 -0.5175  0.5268  0.5458
#> 
#> Columns 19 to 20  0.5998  0.4230
#>   0.6364  0.6883
#>   0.3622  0.6000
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>