RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8211  0.3586  0.0145  0.5269  0.6034  0.4182 -0.4217  0.2249  0.0395
#>   0.7415 -0.8618 -0.7936  0.8103  0.1840  0.5219  0.3983 -0.7840 -0.9605
#>   0.3213  0.5454 -0.6046 -0.0349 -0.3584 -0.2448 -0.2053 -0.8139 -0.7573
#> 
#> Columns 10 to 18 -0.4692  0.0305 -0.7176 -0.0516  0.5812  0.5803  0.6240 -0.2721 -0.0328
#>   0.1734 -0.1763 -0.1478  0.2378 -0.2532 -0.7658 -0.3427 -0.7404  0.6647
#>   0.8237  0.7864 -0.2362 -0.0955  0.4582  0.7829  0.0009  0.8078 -0.7088
#> 
#> Columns 19 to 20 -0.8714  0.4683
#>  -0.8956 -0.6836
#>   0.1868  0.2184
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3546  0.6445 -0.3338 -0.1106 -0.0930 -0.1433  0.1977 -0.3046 -0.6739
#>   0.5304 -0.2793 -0.7060 -0.2306  0.4916  0.4411  0.6123  0.0045  0.2369
#>   0.3946 -0.3366 -0.5137  0.4432  0.3180  0.3244  0.2124 -0.1586 -0.4759
#> 
#> Columns 10 to 18  0.4357  0.4062  0.0900  0.4413 -0.4081  0.3182  0.2818 -0.0838 -0.3819
#>   0.4127 -0.6406  0.2803  0.0547 -0.5325 -0.2941 -0.2360  0.3237 -0.0077
#>  -0.1718 -0.0175  0.7818  0.2801 -0.4825  0.1965 -0.1155  0.4818  0.7558
#> 
#> Columns 19 to 20 -0.3683 -0.0834
#>  -0.1267 -0.7596
#>   0.1546  0.1576
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3870 -0.1729  0.0451  0.3804  0.0574  0.4860 -0.3278 -0.2400 -0.2689
#>   0.4482 -0.2643 -0.5995 -0.1047 -0.2924  0.3297 -0.2621 -0.1724 -0.3043
#>  -0.0665 -0.5025 -0.5849  0.1530 -0.4513  0.1839 -0.2518  0.3089 -0.1259
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1793  0.0904 -0.5049 -0.7367 -0.3171 -0.3479  0.9019  0.2600 -0.6034
#>  -0.5583  0.6419 -0.5297 -0.2035 -0.1203  0.0840 -0.5968 -0.2547  0.6955
#>  -0.4162  0.7741 -0.4963 -0.4793 -0.6053  0.1124 -0.4674 -0.2934  0.0368
#> 
#> Columns 10 to 18  0.1004  0.7227  0.2577 -0.3775  0.6450 -0.4465 -0.5278  0.4435  0.1842
#>   0.5257 -0.6449  0.1598 -0.0984  0.4825 -0.7795  0.6255  0.4569 -0.6095
#>   0.0482  0.3041  0.3801 -0.5053  0.2676 -0.1241  0.0726 -0.6296 -0.4279
#> 
#> Columns 19 to 20  0.3438 -0.2594
#>   0.3386 -0.2748
#>  -0.3946  0.7152
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3075  0.3690 -0.1481 -0.4030 -0.4741 -0.0350  0.0792 -0.4927 -0.6629
#>  -0.0509 -0.1428 -0.5756  0.1857 -0.4353 -0.0123 -0.6027  0.0897 -0.2059
#>   0.0066 -0.2682  0.2456  0.0483 -0.3232  0.0060 -0.6075 -0.2828 -0.2408
#> 
#> Columns 10 to 18  0.5631  0.4666  0.0517  0.0671  0.0823  0.3751 -0.4807  0.2750 -0.6706
#>   0.0144  0.0049 -0.2019 -0.1264  0.7289  0.7006  0.1729  0.6336  0.0629
#>   0.0281 -0.1108  0.4135 -0.0174  0.3991  0.2376 -0.0812  0.3587  0.1736
#> 
#> Columns 19 to 20 -0.5031  0.1861
#>  -0.1228  0.3733
#>  -0.2930  0.0956
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>