RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3510  0.4765  0.3482  0.0342  0.1512  0.4429  0.5678 -0.2549  0.3399
#>  -0.9130  0.4339 -0.7116  0.6765 -0.8284  0.0886 -0.4373  0.7482 -0.1746
#>   0.4688  0.0164 -0.1409 -0.5770  0.1972  0.0658 -0.2626  0.2720  0.2823
#> 
#> Columns 10 to 18  0.1700  0.2597 -0.4225 -0.3061 -0.1911 -0.6963  0.5586  0.0096 -0.8111
#>   0.9388 -0.3643 -0.1247 -0.8574  0.8430 -0.2402 -0.1531 -0.7027 -0.1192
#>   0.6603 -0.0548  0.6962 -0.1416 -0.1877  0.5277  0.6539  0.6118  0.2878
#> 
#> Columns 19 to 20 -0.1063 -0.4574
#>  -0.3148  0.1195
#>   0.5449  0.1347
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4214  0.5062  0.0676 -0.1891 -0.1464 -0.3959  0.5937 -0.1060 -0.2561
#>   0.3073  0.1628  0.4089 -0.3000  0.4387  0.3091  0.0118  0.0194 -0.4005
#>   0.2130  0.1511  0.3660 -0.4661  0.6693 -0.6408  0.2164  0.0590  0.5342
#> 
#> Columns 10 to 18  0.1094  0.2182 -0.1703 -0.2704  0.0886 -0.2054  0.5483 -0.5291 -0.7990
#>   0.0606  0.1133 -0.1601 -0.3311 -0.0906 -0.0273  0.1156  0.3204 -0.2852
#>  -0.5371 -0.1219 -0.3319 -0.0578  0.1638 -0.7189  0.0101  0.1747  0.1079
#> 
#> Columns 19 to 20  0.4183 -0.1350
#>   0.2398 -0.0445
#>   0.2665  0.1087
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1933  0.1272 -0.3507  0.1364 -0.1554  0.0688  0.4575 -0.1747 -0.2355
#>  -0.1158  0.5681 -0.5070 -0.2756  0.1922 -0.4495  0.0828  0.1330  0.0124
#>   0.2402 -0.0066  0.0297 -0.1624  0.0683  0.0393  0.3642 -0.1880  0.2668
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4541  0.2758  0.2265  0.6898  0.3772  0.4710  0.0755 -0.0940  0.1481
#>   0.0153  0.1323  0.3858  0.1070  0.7034  0.7580  0.0266 -0.3020 -0.6181
#>  -0.2144 -0.0324  0.6420  0.1871  0.2451  0.0702 -0.0620  0.1140 -0.2811
#> 
#> Columns 10 to 18  0.1970  0.3415  0.1085  0.1975 -0.1921  0.7488  0.0452 -0.1415 -0.1279
#>  -0.8288  0.2686  0.0503  0.5527 -0.1914 -0.2427  0.2802 -0.0341  0.4861
#>  -0.5848 -0.4863  0.5881  0.6418  0.3347  0.4240 -0.5480  0.3650  0.0045
#> 
#> Columns 19 to 20  0.5217 -0.5777
#>  -0.3185  0.2030
#>  -0.7171 -0.4448
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1980 -0.1358  0.2428  0.1319  0.3272 -0.2321  0.2963 -0.1222 -0.4840
#>  -0.3824  0.3157 -0.0536  0.3420 -0.3089 -0.2214  0.1367 -0.0266 -0.0854
#>  -0.3301  0.0836  0.1808 -0.0087  0.2417 -0.4176  0.1085  0.0487  0.3436
#> 
#> Columns 10 to 18 -0.3932  0.1886 -0.4001 -0.1339  0.3615 -0.3766  0.2412 -0.2945 -0.5717
#>   0.2294  0.0387 -0.4026 -0.1430  0.2543  0.0851  0.1344 -0.6450 -0.3254
#>  -0.0789 -0.0991 -0.3922 -0.6278 -0.1207 -0.4960  0.1862 -0.5618 -0.4378
#> 
#> Columns 19 to 20  0.3848  0.3265
#>   0.1244  0.6534
#>   0.4788  0.5938
#> [ CPUFloatType{2,3,20} ]
#>