RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8532  0.3311  0.9690  0.7724 -0.9007  0.9057  0.2458  0.9128 -0.2698
#>  -0.0811 -0.0427 -0.5107  0.3040  0.0428  0.4134  0.0127  0.1124 -0.1529
#>  -0.4435 -0.5837 -0.1005  0.7706  0.3579 -0.0499  0.1632  0.2653 -0.1685
#> 
#> Columns 10 to 18  0.4016  0.8238  0.4574 -0.6631  0.8320  0.8747 -0.6242  0.5617  0.0755
#>   0.5402 -0.5139 -0.8421  0.6029 -0.2853 -0.7031 -0.8647 -0.8290  0.5502
#>   0.2237 -0.0835  0.1571 -0.7365 -0.7259 -0.0407  0.1251  0.3433 -0.0597
#> 
#> Columns 19 to 20  0.9112  0.0194
#>   0.8988  0.1257
#>   0.5393 -0.0166
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6655  0.3931  0.2768 -0.1221  0.0379  0.6547  0.0814  0.8448 -0.3203
#>   0.4409  0.1893  0.2269  0.1177  0.1312 -0.3811  0.7868 -0.5369 -0.6368
#>   0.4691  0.3445  0.1964 -0.0725  0.3748 -0.3437  0.7613 -0.2594 -0.4716
#> 
#> Columns 10 to 18 -0.0552 -0.1080 -0.5815  0.0669  0.5516  0.2906 -0.0443 -0.7580 -0.2269
#>  -0.1444 -0.3612 -0.6944  0.2651 -0.1986 -0.3913  0.2481 -0.2404  0.0869
#>  -0.4697  0.1910 -0.1963 -0.7141 -0.2552  0.0302  0.2396  0.1965  0.2009
#> 
#> Columns 19 to 20  0.5811  0.3944
#>   0.4583 -0.3651
#>   0.5412  0.4132
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3674  0.1614  0.1410  0.3533 -0.0574  0.2173 -0.1148  0.4907 -0.3443
#>   0.2092 -0.1491  0.6168 -0.0887 -0.4419  0.5675  0.6238  0.5855 -0.1994
#>   0.4664  0.1030  0.1649 -0.0393  0.1728  0.2795  0.1169  0.4400  0.0725
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8188 -0.0050 -0.1443 -0.2424  0.4123 -0.2606  0.0834  0.6012 -0.2582
#>   0.0568  0.5607  0.4007  0.2724  0.5858 -0.7517  0.4104 -0.0773  0.5693
#>   0.1368 -0.5104  0.0898 -0.3817 -0.2080  0.5796 -0.4317 -0.7310 -0.1849
#> 
#> Columns 10 to 18  0.5425 -0.1598  0.7354  0.6832 -0.2540  0.0452 -0.2211  0.3240 -0.7773
#>  -0.3655  0.1860  0.3103 -0.4142  0.3702  0.3355 -0.0492  0.2028 -0.5498
#>   0.2012 -0.5588  0.0562  0.1885  0.5949 -0.0928 -0.0608 -0.6684  0.4290
#> 
#> Columns 19 to 20 -0.1166  0.0775
#>   0.0344 -0.6639
#>   0.4132  0.3383
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4717  0.0766  0.5761  0.0750 -0.2822 -0.1710  0.5168  0.2229 -0.3757
#>   0.3584  0.5024  0.1379  0.7318  0.1438  0.3238  0.2415  0.2952 -0.1783
#>  -0.0978 -0.3396  0.6487 -0.1872  0.3134  0.3506 -0.0452  0.1810 -0.1854
#> 
#> Columns 10 to 18  0.0824  0.0415 -0.2217 -0.1613 -0.0305 -0.0940 -0.2189 -0.3045 -0.2317
#>  -0.3492 -0.3200 -0.0435  0.1010  0.1572 -0.0540 -0.3483 -0.1672  0.4660
#>  -0.5181 -0.1068 -0.2556 -0.2815 -0.0484  0.2441  0.1848 -0.3989 -0.0589
#> 
#> Columns 19 to 20  0.5615 -0.3402
#>   0.7820  0.2860
#>   0.5584 -0.1844
#> [ CPUFloatType{2,3,20} ]
#>