RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5381 -0.9447 -0.1934 -0.5398 -0.0953  0.3949  0.7479  0.4872 -0.5820
#>   0.5978 -0.2441  0.8381 -0.2635  0.2030  0.7390 -0.4227 -0.3739 -0.4695
#>  -0.2902  0.0812  0.5738  0.5072  0.2972  0.4085 -0.7782  0.8607  0.5496
#> 
#> Columns 10 to 18 -0.6470  0.5436  0.8756 -0.4592 -0.4074  0.5575  0.0513  0.1179 -0.6812
#>  -0.1327  0.6514 -0.3260 -0.6852  0.3146  0.5476 -0.2800 -0.3870 -0.2987
#>  -0.7493  0.0129  0.2342 -0.0437  0.1844  0.6656  0.5063 -0.1865 -0.0016
#> 
#> Columns 19 to 20  0.5066 -0.6166
#>   0.6339  0.5123
#>   0.6254  0.1023
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3815  0.1944  0.3105  0.2689  0.1135 -0.3607  0.1202  0.4524 -0.2709
#>   0.5766 -0.3554  0.5782 -0.3764 -0.0491  0.0457  0.2227 -0.0643 -0.3212
#>  -0.1185 -0.0808  0.4842  0.2722  0.2450 -0.0790  0.4617 -0.2810 -0.1641
#> 
#> Columns 10 to 18 -0.4182  0.2034  0.0431 -0.4885 -0.4523  0.2243  0.4460 -0.0529 -0.2102
#>  -0.0923  0.1430 -0.3680 -0.0218  0.1343  0.1439 -0.2547 -0.3116 -0.3422
#>   0.1666  0.0241 -0.4393 -0.4811 -0.1631 -0.0385  0.0836 -0.0755  0.0088
#> 
#> Columns 19 to 20  0.3617  0.6192
#>   0.5414  0.0146
#>   0.6505 -0.4319
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3553  0.1046  0.0066 -0.0253  0.5054 -0.0445  0.2716 -0.0555 -0.4196
#>   0.5548 -0.3136  0.4677 -0.1930 -0.3472  0.4264  0.3414  0.6876  0.2547
#>   0.6517  0.4334  0.3026  0.1747 -0.2671  0.2331  0.3823  0.4662  0.1315
#> 
#> Columns 10 to 18 -0.2565 -0.1360  0.0339 -0.3679 -0.3182  0.1565  0.4847 -0.3234 -0.1224
#>  -0.0296 -0.1420 -0.2624 -0.2111  0.2021 -0.1374  0.3067 -0.1566 -0.7294
#>  -0.4467  0.2668 -0.0308 -0.6109  0.1798 -0.1045  0.0368 -0.0221  0.1857
#> 
#> Columns 19 to 20  0.7213 -0.1463
#>   0.0844  0.0657
#>   0.6805  0.7073
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.2383 -0.0705  0.2275  0.0833  0.1531  0.1800  0.0576  0.3613 -0.1542
#>   0.6663  0.5418  0.0184  0.4669 -0.1120  0.4904  0.0950  0.7718  0.2887
#>   0.3820 -0.0047  0.3620  0.2056  0.4246  0.5457  0.0734  0.4043 -0.4492
#> 
#> Columns 10 to 18 -0.4275 -0.0976 -0.2639 -0.3699 -0.2074  0.2547  0.3775 -0.3350 -0.2868
#>  -0.4421 -0.1765 -0.3304  0.0495  0.2341 -0.2772  0.3005 -0.3932 -0.2173
#>  -0.2908 -0.2977 -0.6416  0.0583  0.1220  0.2481  0.0776 -0.6995 -0.1041
#> 
#> Columns 19 to 20  0.5312  0.2638
#>   0.6339  0.4734
#>   0.4828 -0.2540
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.3045  0.0111  0.1063 -0.1047  0.3590  0.1608  0.2544 -0.0251 -0.5545
#>   0.2467  0.2321  0.4442  0.5468  0.3108  0.3717  0.2824  0.0565 -0.4422
#>   0.5415  0.2534 -0.0047  0.0335 -0.3466 -0.0731  0.3918  0.3128 -0.0129
#> 
#> Columns 10 to 18 -0.3888 -0.2852 -0.4624 -0.0884 -0.1428 -0.1111  0.3965 -0.4110 -0.2189
#>  -0.5126 -0.1850 -0.5425  0.4943  0.2796  0.2314  0.0015 -0.4928  0.4692
#>  -0.1770 -0.1214 -0.1516  0.0268 -0.1394 -0.2936  0.3330 -0.2900 -0.3666
#> 
#> Columns 19 to 20  0.5958 -0.4008
#>   0.7263 -0.2043
#>   0.6073  0.1662
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.0476 -0.6134  0.2624  0.5686 -0.5277  0.2725 -0.3884  0.0101  0.1092
#>   0.4758 -0.5602  0.0113  0.6405 -0.1315  0.1682  0.0760 -0.7843  0.6107
#>  -0.5122  0.3335 -0.6738 -0.3274  0.4695 -0.4652  0.3587  0.2126 -0.6800
#> 
#> Columns 10 to 18  0.7676 -0.0532  0.6245  0.1963  0.2383 -0.2156 -0.0663  0.1530  0.1449
#>   0.6808  0.1250 -0.2490 -0.1872  0.3152  0.4321 -0.2201 -0.4692 -0.0682
#>  -0.3883  0.3170  0.5747  0.8250 -0.7089 -0.1686 -0.8279  0.2349 -0.3895
#> 
#> Columns 19 to 20 -0.2681  0.5747
#>   0.1374  0.5855
#>  -0.2335 -0.4813
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3045  0.0111  0.1063 -0.1047  0.3590  0.1608  0.2544 -0.0251 -0.5545
#>   0.2467  0.2321  0.4442  0.5468  0.3108  0.3717  0.2824  0.0565 -0.4422
#>   0.5415  0.2534 -0.0047  0.0335 -0.3466 -0.0731  0.3918  0.3128 -0.0129
#> 
#> Columns 10 to 18 -0.3888 -0.2852 -0.4624 -0.0884 -0.1428 -0.1111  0.3965 -0.4110 -0.2189
#>  -0.5126 -0.1850 -0.5425  0.4943  0.2796  0.2314  0.0015 -0.4928  0.4692
#>  -0.1770 -0.1214 -0.1516  0.0268 -0.1394 -0.2936  0.3330 -0.2900 -0.3666
#> 
#> Columns 19 to 20  0.5958 -0.4008
#>   0.7263 -0.2043
#>   0.6073  0.1662
#> [ CPUFloatType{2,3,20} ]
#>