RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.7275 -0.6475  0.7453  0.7422 -0.5010  0.6163  0.2457  0.8652  0.1801
#>   0.0267 -0.1462  0.4417 -0.0183  0.4304 -0.2282 -0.1126  0.2382  0.4926
#>  -0.7907 -0.2923 -0.6332  0.5934  0.3488 -0.0553  0.2800  0.5252  0.2598
#> 
#> Columns 10 to 18  0.6055  0.1640  0.6420 -0.2776  0.4851  0.1263 -0.4616 -0.3454 -0.6760
#>  -0.9141 -0.0076  0.0872  0.7751 -0.0902  0.2994  0.7807 -0.8250 -0.1828
#>   0.5521 -0.2259 -0.1091  0.3216 -0.1968 -0.4909  0.7564 -0.2136 -0.2050
#> 
#> Columns 19 to 20 -0.4152 -0.6146
#>  -0.6749 -0.2228
#>   0.3535 -0.8388
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6104  0.6321  0.1265  0.5081 -0.1982 -0.2641  0.4268  0.7190 -0.1548
#>  -0.2471 -0.2347  0.4291  0.2903  0.4130 -0.3172  0.1214 -0.0699 -0.1897
#>  -0.0914  0.4979  0.2330  0.4960  0.0254 -0.0194  0.6907  0.2556 -0.0957
#> 
#> Columns 10 to 18 -0.5004 -0.3034  0.0547  0.1371  0.2896 -0.1140 -0.0601 -0.0319  0.0638
#>   0.2054 -0.2059 -0.3275 -0.2327 -0.3263 -0.1894  0.3962 -0.3744 -0.4840
#>  -0.6387 -0.2660 -0.3831  0.1326  0.4679 -0.1538  0.3019 -0.0119  0.2045
#> 
#> Columns 19 to 20  0.7418  0.4546
#>   0.5061 -0.6578
#>  -0.1669  0.3774
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2245  0.2071  0.5083  0.3152  0.1962  0.5919  0.2038  0.4792 -0.3146
#>   0.4256  0.3291  0.3462  0.0885  0.0712 -0.0500  0.2194  0.1502 -0.1683
#>   0.1078  0.3053  0.1272  0.6888  0.1202  0.5951  0.5984  0.3777 -0.3873
#> 
#> Columns 10 to 18 -0.2876 -0.0273 -0.1564  0.0109  0.0003  0.1969  0.0574 -0.6259  0.2561
#>  -0.6542 -0.1590 -0.1491 -0.3995  0.1430 -0.0672  0.2458 -0.0199 -0.0084
#>   0.0256 -0.4715 -0.3951  0.0626  0.2625 -0.1129  0.2223 -0.3179 -0.0886
#> 
#> Columns 19 to 20  0.6942 -0.2465
#>   0.5929  0.2285
#>   0.4106 -0.0381
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.1747 -0.0439  0.4421 -0.1997  0.1320  0.4048 -0.2865  0.2917 -0.0738
#>   0.2748  0.1596  0.5073  0.2709  0.0739  0.2917  0.5614  0.3368 -0.1689
#>   0.4759  0.5596 -0.1025  0.5117 -0.1903  0.4824  0.4359  0.4491 -0.1111
#> 
#> Columns 10 to 18 -0.5789 -0.3782 -0.4966 -0.2788  0.0596 -0.4176 -0.1725 -0.6924 -0.0285
#>  -0.1134  0.0924 -0.0714 -0.1223 -0.2166  0.2994  0.2253 -0.1980 -0.2669
#>  -0.5011 -0.3455 -0.2124 -0.0527  0.2533 -0.2188  0.0176 -0.2045  0.0966
#> 
#> Columns 19 to 20  0.5039 -0.0722
#>   0.7000  0.1774
#>   0.6483  0.5409
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.3904 -0.1434  0.4421 -0.0112  0.0157  0.0531  0.3288  0.2608 -0.2212
#>   0.2333 -0.1860  0.4021  0.1414  0.0560  0.4052  0.0933  0.0300 -0.3218
#>   0.4334  0.2935  0.1888  0.2269  0.0426  0.6096  0.3379  0.6168 -0.2407
#> 
#> Columns 10 to 18 -0.2409 -0.0475 -0.5055  0.0344 -0.0278 -0.2599  0.0237 -0.4140 -0.2875
#>  -0.4442 -0.2534 -0.3233 -0.5271  0.1127 -0.1166 -0.3038 -0.1466  0.3550
#>  -0.2823 -0.1752 -0.5050  0.2095  0.2502 -0.0020  0.1925 -0.6283 -0.2484
#> 
#> Columns 19 to 20  0.6138 -0.0721
#>   0.4809 -0.0288
#>   0.6176  0.0201
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.1404 -0.3722 -0.3551 -0.1027 -0.0032  0.1066 -0.1100 -0.2829 -0.1218
#>   0.1587 -0.1236  0.3103 -0.4491 -0.7017  0.6694 -0.5886 -0.2659 -0.0929
#>   0.1268 -0.3704 -0.7322  0.8279  0.7387  0.3583  0.3909 -0.3171  0.1649
#> 
#> Columns 10 to 18  0.0924 -0.4063  0.4337 -0.0431  0.1865 -0.0735 -0.4041 -0.0101 -0.1795
#>   0.7374 -0.3695 -0.5202  0.1805 -0.0620 -0.1016  0.3405 -0.4985  0.6591
#>  -0.5181 -0.3755 -0.0237  0.3097 -0.5156 -0.5241 -0.6976 -0.2120 -0.5876
#> 
#> Columns 19 to 20 -0.3297 -0.1023
#>   0.0534  0.1124
#>  -0.4134  0.2033
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3904 -0.1434  0.4421 -0.0112  0.0157  0.0531  0.3288  0.2608 -0.2212
#>   0.2333 -0.1860  0.4021  0.1414  0.0560  0.4052  0.0933  0.0300 -0.3218
#>   0.4334  0.2935  0.1888  0.2269  0.0426  0.6096  0.3379  0.6168 -0.2407
#> 
#> Columns 10 to 18 -0.2409 -0.0475 -0.5055  0.0344 -0.0278 -0.2599  0.0237 -0.4140 -0.2875
#>  -0.4442 -0.2534 -0.3233 -0.5271  0.1127 -0.1166 -0.3038 -0.1466  0.3550
#>  -0.2823 -0.1752 -0.5050  0.2095  0.2502 -0.0020  0.1925 -0.6283 -0.2484
#> 
#> Columns 19 to 20  0.6138 -0.0721
#>   0.4809 -0.0288
#>   0.6176  0.0201
#> [ CPUFloatType{2,3,20} ]
#>