RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6248 -0.7798 -0.5184 -0.9696  0.3317  0.0534  0.6321 -0.2086 -0.0529
#>   0.0883 -0.3858 -0.1922 -0.5654 -0.5233  0.0305  0.4492 -0.6662 -0.2639
#>   0.3463  0.1557  0.4640 -0.2528  0.2095 -0.4346  0.8649  0.3558 -0.8582
#> 
#> Columns 10 to 18  0.3266  0.0047  0.2434 -0.0704 -0.7736 -0.4887  0.8323 -0.0679 -0.7220
#>  -0.4067 -0.4481 -0.2739 -0.1924  0.6102 -0.7748 -0.1373 -0.2178  0.3279
#>   0.9133 -0.3228 -0.2992 -0.8183 -0.9501 -0.9195  0.5305 -0.3727 -0.1818
#> 
#> Columns 19 to 20 -0.4770  0.0042
#>  -0.6854  0.4565
#>   0.1206 -0.8135
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4701  0.5429 -0.1449 -0.9259  0.2893 -0.1816  0.4904 -0.0544  0.5019
#>  -0.2365  0.2675 -0.0401 -0.4440 -0.4569 -0.3045 -0.0156 -0.0885  0.3523
#>  -0.7451 -0.4677  0.2148 -0.7111 -0.1977 -0.1749  0.2448  0.3339 -0.3400
#> 
#> Columns 10 to 18  0.8289 -0.3381 -0.1734 -0.0182 -0.2842 -0.4357 -0.7186 -0.6717  0.3302
#>   0.2231 -0.4881 -0.5683 -0.5203 -0.2549 -0.4247  0.0165  0.2711  0.3018
#>   0.5007 -0.2599 -0.7920  0.1322 -0.4315 -0.4765 -0.5726 -0.1104  0.1514
#> 
#> Columns 19 to 20 -0.1894 -0.5145
#>  -0.7560  0.1620
#>  -0.3517  0.1173
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5197 -0.2220 -0.4468 -0.0134 -0.3154 -0.0536  0.4460  0.4094 -0.1215
#>  -0.2590 -0.1998 -0.0034 -0.5033 -0.1713 -0.3639 -0.2351  0.2928  0.3541
#>   0.1329  0.1156 -0.4441 -0.3680 -0.0464  0.0890  0.6832  0.4949  0.1813
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0248  0.5965  0.3128  0.3513 -0.2354  0.4143 -0.4996 -0.4406 -0.8894
#>  -0.1812 -0.6415 -0.3426 -0.0430  0.5288 -0.1629  0.4040  0.0285  0.1078
#>  -0.1788 -0.5767 -0.5911 -0.4414  0.6532 -0.4962  0.2378  0.2195  0.0232
#> 
#> Columns 10 to 18 -0.8070  0.3883  0.5652 -0.2486 -0.0582 -0.4310  0.2835 -0.4383  0.2371
#>  -0.2495  0.3649  0.5674 -0.5681  0.4128 -0.6556  0.6390  0.7492  0.2271
#>   0.0977 -0.3286 -0.0227 -0.7634 -0.4299  0.4978  0.6818  0.0096 -0.0775
#> 
#> Columns 19 to 20  0.3649  0.2064
#>   0.3774 -0.1598
#>   0.4547  0.7573
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2328  0.0893 -0.2665 -0.4768 -0.1689 -0.1267  0.0794  0.6362  0.3512
#>  -0.0959  0.4851 -0.1506 -0.6131  0.3466  0.1850  0.3749 -0.2182  0.1221
#>  -0.5225 -0.2310 -0.0414 -0.5458 -0.2149  0.0554  0.2121  0.0971 -0.3865
#> 
#> Columns 10 to 18  0.1444 -0.5022 -0.4931  0.2208 -0.0443  0.3751 -0.3645  0.1599 -0.1663
#>   0.7723 -0.3548 -0.0539  0.1450 -0.5726 -0.4471 -0.5273  0.0552  0.1551
#>   0.5769 -0.1698 -0.2729  0.5055 -0.4903 -0.4020 -0.6431 -0.1563  0.2647
#> 
#> Columns 19 to 20  0.1798 -0.2684
#>  -0.0650 -0.5713
#>  -0.2316 -0.0257
#> [ CPUFloatType{2,3,20} ]
#>