RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2472 -0.4901  0.1519 -0.4664  0.8803  0.6010 -0.4348 -0.1186 -0.6441
#>  -0.9832  0.2091  0.4752  0.2244  0.4922 -0.4428 -0.6684  0.1091 -0.5077
#>  -0.4957 -0.7452  0.5911 -0.8353  0.3810 -0.7685 -0.8642 -0.5195 -0.3920
#> 
#> Columns 10 to 18  0.5172 -0.6801 -0.8373 -0.3628  0.5421 -0.4041 -0.5078 -0.6134  0.0687
#>   0.7342  0.1057 -0.5948  0.0821  0.8351 -0.6557  0.6396  0.2889  0.7071
#>  -0.0793 -0.5525 -0.1251 -0.6909  0.5982 -0.4656  0.1234  0.2594 -0.7304
#> 
#> Columns 19 to 20  0.1716 -0.0052
#>   0.0959 -0.5883
#>   0.6681 -0.0196
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.7206  0.1656 -0.1608 -0.6311  0.4182  0.1041 -0.1818  0.0233 -0.2239
#>  -0.1717 -0.5278  0.4450 -0.4830  0.3259  0.3383 -0.0175 -0.2852 -0.8208
#>  -0.6416  0.2701 -0.5767 -0.5135  0.5569  0.0893 -0.2262 -0.4333 -0.0486
#> 
#> Columns 10 to 18 -0.3261 -0.3235 -0.5837 -0.5588  0.7849  0.0807  0.7510 -0.0173  0.2747
#>  -0.2873 -0.5204 -0.5659 -0.3163  0.2890  0.0356  0.4201  0.2349 -0.2221
#>  -0.4169 -0.0208 -0.3796 -0.6362  0.6431  0.1128  0.6868 -0.1403  0.1990
#> 
#> Columns 19 to 20  0.3773 -0.2797
#>  -0.4193  0.3656
#>   0.1533  0.4187
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5009 -0.4169 -0.1166 -0.5552  0.2252  0.1934 -0.2756 -0.2281 -0.2904
#>  -0.7351  0.5469 -0.4084 -0.2108  0.5283 -0.1976 -0.0405 -0.3487 -0.2428
#>  -0.3887 -0.2534 -0.7356 -0.6270  0.5146 -0.1146 -0.1085  0.0368  0.1837
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7495 -0.1178  0.2269 -0.7001 -0.5209 -0.2053  0.1766  0.0339  0.0303
#>   0.4290  0.1589  0.3707 -0.4123  0.6129  0.3684 -0.7305  0.1349 -0.4921
#>   0.0763  0.0470 -0.2569 -0.0570 -0.0983  0.2015 -0.3411 -0.4605 -0.0744
#> 
#> Columns 10 to 18  0.1151 -0.4592  0.2412  0.5628 -0.0367 -0.1217 -0.7444  0.1585  0.6518
#>  -0.1846 -0.3985  0.6245  0.1885 -0.8240 -0.4949 -0.3246 -0.0762 -0.2753
#>  -0.0917  0.3332  0.2765 -0.0337 -0.3606  0.0414 -0.1346  0.0034  0.3043
#> 
#> Columns 19 to 20 -0.3152  0.2280
#>   0.2055  0.6799
#>   0.3150 -0.4449
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4937  0.0014 -0.3564 -0.2296  0.3348  0.0798 -0.0162 -0.0980 -0.1380
#>  -0.2945  0.1354 -0.3791 -0.6591  0.1464 -0.3036 -0.1283 -0.1452 -0.0183
#>  -0.3935 -0.3162 -0.3958 -0.6707  0.5224 -0.3850 -0.0854 -0.0516 -0.1815
#> 
#> Columns 10 to 18  0.2121  0.2727 -0.4570 -0.5735  0.5362  0.5833  0.5442 -0.3879  0.0125
#>  -0.4750  0.4191 -0.2160 -0.4994  0.5351  0.5541  0.7911 -0.0400 -0.1670
#>  -0.4816  0.1980 -0.3086 -0.4065  0.4029  0.3524  0.5604 -0.1306 -0.1867
#> 
#> Columns 19 to 20  0.3994 -0.2271
#>   0.4265 -0.4606
#>   0.3718 -0.0780
#> [ CPUFloatType{2,3,20} ]
#>