RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1624  0.2949  0.5036 -0.7665 -0.6221 -0.3223 -0.7286  0.6707  0.3304
#>   0.0156  0.6097 -0.4405  0.1229 -0.8694 -0.7723 -0.0948  0.7826  0.3100
#>   0.5264  0.6840  0.0139 -0.2076  0.3473  0.7837 -0.3314 -0.4092  0.1674
#> 
#> Columns 10 to 18  0.6049  0.2750  0.3854 -0.1880  0.0139 -0.1690  0.0058 -0.4393 -0.4280
#>   0.7418  0.5586 -0.5220  0.5731 -0.4051 -0.4480 -0.7307 -0.1937  0.1530
#>  -0.8536  0.5681 -0.3704  0.0664 -0.4614 -0.1683 -0.1723 -0.3615 -0.5477
#> 
#> Columns 19 to 20 -0.4196 -0.0088
#>  -0.6877  0.5943
#>   0.7660  0.8219
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1531  0.4248  0.8312 -0.1533  0.0178 -0.3912 -0.2959 -0.0652 -0.4093
#>   0.3135 -0.0999  0.4056 -0.6223 -0.0655 -0.6209  0.5742 -0.0167 -0.3901
#>   0.1495 -0.2108 -0.0642 -0.0928 -0.0809  0.1699  0.1751  0.0453  0.0353
#> 
#> Columns 10 to 18  0.8162 -0.2141  0.2969  0.5760 -0.3900  0.1548  0.6057 -0.1473  0.1801
#>   0.4882 -0.2257  0.2430  0.7847  0.2161 -0.3962  0.5716 -0.1647  0.3384
#>  -0.4545 -0.5955 -0.1492  0.0750  0.6026  0.2926 -0.0924  0.2614  0.0722
#> 
#> Columns 19 to 20  0.0700  0.4985
#>  -0.2183  0.6709
#>   0.4942 -0.2742
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0188  0.3760 -0.1237 -0.0465  0.1602  0.1865  0.1879  0.3171 -0.2457
#>  -0.3048 -0.1035  0.2996 -0.2165 -0.1155 -0.0211 -0.0210 -0.4782 -0.2402
#>  -0.1644  0.4043 -0.2555 -0.2934 -0.0068  0.3403 -0.5388 -0.0999  0.0056
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4062  0.3475 -0.4835  0.6746 -0.7281  0.5174 -0.5109 -0.0664  0.3129
#>   0.4659  0.3600  0.2362  0.6887 -0.0759  0.3320 -0.4879 -0.0439 -0.3903
#>  -0.5033  0.5639 -0.1024 -0.2806  0.1541 -0.2824  0.2963  0.1543  0.4110
#> 
#> Columns 10 to 18  0.1934 -0.7449  0.2897  0.5228  0.0521 -0.4219 -0.0846  0.2511  0.4832
#>   0.0797 -0.5721 -0.4146  0.1215  0.0777 -0.3569  0.2438 -0.1152  0.3274
#>  -0.1163  0.0821  0.1666  0.2220  0.0485  0.0278  0.1548 -0.4355  0.5065
#> 
#> Columns 19 to 20  0.2331  0.4491
#>   0.0715  0.2723
#>   0.2243  0.3297
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5404  0.1109  0.1514 -0.3559 -0.0264 -0.3133  0.2058 -0.0314 -0.5009
#>  -0.2350 -0.0804 -0.0510 -0.3782 -0.3146 -0.3643  0.4078  0.1541 -0.3994
#>   0.0383  0.1639  0.1081 -0.5680 -0.3290  0.1913 -0.4608  0.2360  0.0206
#> 
#> Columns 10 to 18 -0.4289  0.4242 -0.1151  0.5977  0.0862  0.0905  0.0702 -0.3430 -0.0996
#>  -0.1998  0.0802 -0.2608  0.4619 -0.3019  0.1871 -0.1047  0.1391 -0.1166
#>   0.4552  0.4017  0.2573  0.3397 -0.0423 -0.2756  0.2250  0.1432  0.1089
#> 
#> Columns 19 to 20  0.3821  0.1438
#>   0.0205  0.2820
#>   0.1266  0.2602
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>