RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4634  0.7554  0.4191 -0.7432 -0.5622  0.5072  0.2543  0.2990 -0.7158
#>  -0.2867  0.1713  0.3336 -0.4058  0.2110  0.6371 -0.7879  0.6267 -0.0937
#>   0.5145 -0.1362 -0.0043 -0.9158  0.5408 -0.4221  0.4900  0.8114 -0.3184
#> 
#> Columns 10 to 18  0.2553  0.1381 -0.6600 -0.4159 -0.0502 -0.0559 -0.3188  0.7070 -0.7823
#>  -0.8672  0.5129 -0.8364  0.2275 -0.6325  0.4788  0.2747 -0.1632 -0.7000
#>  -0.5963 -0.2925 -0.7142  0.7964  0.6363  0.3452  0.3026 -0.3544  0.0130
#> 
#> Columns 19 to 20 -0.1236 -0.0739
#>   0.1559  0.5853
#>   0.4587  0.2271
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2965  0.1822 -0.0848 -0.0648 -0.0639  0.3841 -0.6067  0.5300 -0.3640
#>   0.0946  0.3883 -0.0216 -0.1036  0.3332  0.4149 -0.2540  0.6602 -0.4813
#>   0.2141  0.2214  0.3776 -0.1884  0.0443 -0.2150  0.4141  0.3904 -0.1403
#> 
#> Columns 10 to 18 -0.1760  0.1331  0.0446  0.5322 -0.1688 -0.0028  0.3360  0.5488  0.1216
#>   0.1183  0.1623 -0.6361  0.2920  0.5316 -0.0921  0.1670  0.2694 -0.5223
#>  -0.2576  0.4660 -0.4819  0.1037  0.5595 -0.1727 -0.2613  0.4837 -0.7604
#> 
#> Columns 19 to 20 -0.2721 -0.6877
#>  -0.0921 -0.2750
#>  -0.2419 -0.1463
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.6763  0.0309  0.1193 -0.7246 -0.1960 -0.0250  0.0606  0.2486  0.1058
#>  -0.0913  0.0004 -0.0563 -0.0051  0.0452  0.4746 -0.2566  0.6096 -0.3740
#>   0.1907  0.3992  0.0395  0.0691  0.2291  0.3381  0.2725  0.3840 -0.1613
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5771  0.4391 -0.1250 -0.3749  0.2538  0.0176  0.0112  0.1197  0.3031
#>   0.0596  0.1827 -0.1081 -0.1906  0.1574  0.5687 -0.2547 -0.4454  0.7222
#>  -0.2779  0.2987  0.6102 -0.6675 -0.3572 -0.5214  0.0802  0.3828  0.1394
#> 
#> Columns 10 to 18 -0.5437 -0.5231 -0.3661  0.0633  0.0118  0.4373 -0.8195  0.0709 -0.4342
#>  -0.1835 -0.6777 -0.7804  0.4126  0.3514 -0.2471 -0.4522  0.1494  0.2147
#>  -0.0267 -0.3383 -0.0360 -0.0320  0.4384  0.2120 -0.7501  0.1490  0.0842
#> 
#> Columns 19 to 20 -0.4147 -0.4691
#>  -0.1935  0.0167
#>  -0.3633 -0.9322
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4940  0.1674  0.2327 -0.3925  0.0693  0.4657 -0.4190  0.0794  0.0163
#>   0.2231 -0.2021  0.0634  0.0543  0.1433  0.2479 -0.3220  0.3887 -0.1913
#>   0.3803  0.3405  0.2277  0.0527  0.0977  0.5997 -0.1646  0.0744  0.0006
#> 
#> Columns 10 to 18  0.4900 -0.4230 -0.2763 -0.2209  0.3998  0.4674  0.3352 -0.0712  0.0726
#>   0.0516  0.2143 -0.2747 -0.0092  0.0491  0.0228  0.0837  0.5096  0.3773
#>   0.7147 -0.2827  0.0350 -0.0938  0.2487  0.2902 -0.0367  0.3583  0.2486
#> 
#> Columns 19 to 20 -0.2668 -0.0347
#>  -0.0867 -0.0962
#>  -0.4083 -0.4808
#> [ CPUFloatType{2,3,20} ]
#>