RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4690  0.5458  0.4642 -0.7454 -0.5733 -0.3449  0.0904  0.4648 -0.0409
#>   0.2905  0.6720 -0.3533 -0.8008  0.7350  0.7263  0.3954  0.2388  0.9157
#>   0.2515 -0.3270  0.2584  0.7347  0.0061 -0.6726  0.8113  0.6573  0.5168
#> 
#> Columns 10 to 18  0.8080 -0.4090 -0.3773 -0.6903  0.6709 -0.4981  0.1650  0.5108 -0.1961
#>  -0.8566 -0.2064  0.1766  0.6224  0.5254  0.0498 -0.0743 -0.6656  0.0992
#>  -0.5087  0.7899  0.8014 -0.4966 -0.6744 -0.4855  0.7990 -0.0324 -0.3314
#> 
#> Columns 19 to 20 -0.3093  0.3809
#>   0.5287 -0.2027
#>   0.5058 -0.0435
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4005 -0.1189 -0.0915  0.2917  0.1626  0.2562  0.0561  0.4001  0.1436
#>  -0.1749 -0.1378  0.0816  0.2795  0.0915 -0.3901  0.5365  0.1222  0.5679
#>  -0.3937 -0.1715  0.0536 -0.3283  0.1035 -0.4139  0.5078 -0.2995 -0.1488
#> 
#> Columns 10 to 18 -0.0777  0.2807  0.3165 -0.0342  0.1101 -0.4689  0.4070 -0.4210  0.4835
#>  -0.4187 -0.4666  0.1701  0.6385 -0.2827 -0.3733  0.3297 -0.4689  0.0959
#>  -0.2810 -0.1748  0.3339 -0.5939  0.3539  0.4269  0.2324 -0.0749  0.2471
#> 
#> Columns 19 to 20  0.0391  0.0519
#>  -0.6386 -0.1334
#>  -0.2513 -0.6207
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0603  0.2352  0.0562 -0.3723 -0.0966 -0.0389  0.2156 -0.2011  0.4335
#>  -0.2975  0.0115  0.2934 -0.1685  0.1014 -0.1888  0.7017  0.3272  0.5133
#>   0.1484  0.2571 -0.0076  0.0547 -0.4005 -0.1566 -0.2585  0.6959  0.2279
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4734 -0.5664  0.2685  0.5274 -0.2625  0.2760 -0.3287 -0.2957 -0.1676
#>   0.4622 -0.3308  0.0225  0.7283 -0.5023  0.1390  0.4963 -0.1821  0.9341
#>  -0.4031 -0.0151  0.5856  0.7356 -0.5216 -0.3064  0.3856  0.1086 -0.5796
#> 
#> Columns 10 to 18 -0.1599  0.3969 -0.6361  0.4961 -0.2306 -0.1049 -0.0851 -0.6309 -0.0433
#>  -0.3519  0.5046  0.3771 -0.4018  0.6936 -0.1947  0.5763 -0.1763 -0.1986
#>   0.6168  0.2535 -0.6797  0.6677  0.2121  0.7241 -0.5623 -0.7394 -0.6940
#> 
#> Columns 19 to 20 -0.5418  0.3276
#>   0.5645  0.3998
#>  -0.1961 -0.2321
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1491 -0.2326  0.1064 -0.3696  0.1578  0.2142 -0.4085  0.6225  0.3649
#>  -0.2953 -0.4307  0.1716 -0.0451  0.1998 -0.0640  0.6591  0.7482  0.5002
#>   0.1197 -0.3147  0.2622 -0.2410 -0.1877 -0.2655 -0.0712  0.2624 -0.1189
#> 
#> Columns 10 to 18 -0.0468  0.0033  0.1769  0.6160 -0.2206  0.0368  0.3308 -0.2886  0.1200
#>   0.2367 -0.0287 -0.0325  0.3517  0.1319  0.0823  0.3725  0.4673  0.3847
#>   0.1812  0.1210  0.4341  0.4111 -0.2892 -0.3940  0.2464 -0.3579  0.2160
#> 
#> Columns 19 to 20  0.1963  0.2960
#>  -0.3452 -0.2643
#>   0.1051  0.3656
#> [ CPUFloatType{2,3,20} ]
#>