RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0518 -0.4217  0.4860  0.9178 -0.2580  0.1085  0.0902  0.5798 -0.1358
#>  -0.3588  0.7367 -0.2517 -0.0327 -0.7808  0.4120 -0.0435  0.9116 -0.7797
#>   0.1022 -0.5056  0.7558 -0.3044 -0.5644  0.0612 -0.2244  0.7798  0.4757
#> 
#> Columns 10 to 18  0.7956  0.4792 -0.1298 -0.8644  0.3067  0.3492  0.0267 -0.0938  0.0312
#>  -0.5032 -0.2691 -0.2006  0.6846  0.7750 -0.1514 -0.4458  0.2850  0.9092
#>   0.1527  0.9069  0.5743 -0.7202  0.4475  0.4895  0.5663 -0.2124 -0.7161
#> 
#> Columns 19 to 20  0.7343 -0.5524
#>   0.5981  0.3453
#>  -0.0234 -0.2192
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2561  0.5106  0.5363  0.1888 -0.0015  0.0755  0.4811  0.5346 -0.2863
#>   0.1512  0.1932  0.2600  0.5634 -0.1147  0.6512  0.6793  0.1563 -0.3199
#>   0.3050 -0.1150  0.5762 -0.0637  0.1199  0.1666  0.1979  0.6365 -0.2250
#> 
#> Columns 10 to 18 -0.7455 -0.5637 -0.5544 -0.1834  0.5786 -0.0463  0.1076 -0.0004  0.0417
#>   0.5292 -0.1241 -0.1867  0.2029 -0.1035 -0.2472 -0.4133 -0.6329 -0.0176
#>  -0.1633 -0.1628 -0.7303  0.2725  0.4009  0.2665  0.4556 -0.6506 -0.5164
#> 
#> Columns 19 to 20  0.3830  0.3879
#>   0.1045 -0.1643
#>   0.2126 -0.1301
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1916 -0.0030  0.1354  0.4893 -0.0401  0.7810  0.2218  0.4788  0.0299
#>   0.4546  0.4521  0.1942  0.4629 -0.5044  0.2026  0.4178  0.5388 -0.3820
#>   0.4787 -0.0096  0.1258  0.2705  0.0132  0.4595 -0.0200  0.3885  0.0339
#> 
#> Columns 10 to 18 -0.3044 -0.5622 -0.1704 -0.3502  0.1541 -0.3048 -0.0351 -0.2818 -0.2820
#>  -0.4719 -0.4825 -0.3237  0.0897  0.5759 -0.5454 -0.2654 -0.2182  0.4177
#>  -0.5443 -0.5510 -0.3812  0.0844  0.2489 -0.2993  0.3267 -0.3404 -0.3641
#> 
#> Columns 19 to 20  0.4559 -0.3423
#>   0.6137  0.3976
#>   0.2902 -0.3125
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.5431  0.5663 -0.0416  0.3163 -0.2574  0.2342  0.2307  0.5718 -0.1873
#>   0.4125  0.2663  0.3543  0.7384 -0.1092  0.6686  0.5028  0.5853 -0.0544
#>   0.3172  0.2004  0.3565  0.1216 -0.2518  0.6568  0.2497  0.7992  0.0509
#> 
#> Columns 10 to 18 -0.4873 -0.1525 -0.2891 -0.1758  0.1157 -0.3164 -0.0675 -0.2979  0.0365
#>  -0.1325 -0.2325 -0.3086  0.1321  0.3195 -0.0979 -0.3378 -0.3699  0.2760
#>  -0.4378 -0.1741 -0.4710  0.1619  0.4762 -0.1849  0.2029 -0.3344 -0.4434
#> 
#> Columns 19 to 20  0.7794  0.2176
#>   0.6369  0.5395
#>   0.3878  0.1410
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.4146  0.3256  0.4410  0.5379  0.0878  0.0924  0.5187  0.2680 -0.5902
#>   0.1226  0.2133  0.2764  0.6660  0.1768  0.5144  0.3265  0.2039 -0.4193
#>   0.4709  0.4793  0.3231  0.5888 -0.2257  0.4099  0.0385  0.4213  0.0326
#> 
#> Columns 10 to 18 -0.3078 -0.1763 -0.4603  0.4168  0.1521  0.2659  0.2038 -0.4614  0.1558
#>  -0.5721 -0.5399 -0.5057  0.0872  0.4876 -0.3598 -0.2393 -0.2198  0.5444
#>  -0.3405 -0.3940 -0.4142  0.3461  0.2625 -0.3638 -0.0591 -0.5737  0.1720
#> 
#> Columns 19 to 20  0.7953 -0.1105
#>   0.6830  0.0563
#>   0.6386 -0.0738
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.1533 -0.3413 -0.5021  0.5479 -0.1348 -0.3177  0.3903 -0.7765  0.6463
#>   0.3016 -0.0141 -0.3892  0.7316 -0.1006  0.5022 -0.2813 -0.5322  0.6094
#>   0.0892 -0.1509 -0.5004  0.6675  0.5968 -0.1042 -0.1861 -0.0327 -0.2679
#> 
#> Columns 10 to 18  0.5839  0.3041  0.3367  0.1459  0.4323  0.5728 -0.6464 -0.0537 -0.4453
#>   0.6286  0.0956 -0.0722 -0.6524  0.2287 -0.0668 -0.2239 -0.4826  0.2780
#>   0.3588 -0.0751 -0.2663  0.2665 -0.2744  0.1342 -0.3854 -0.2005 -0.1187
#> 
#> Columns 19 to 20 -0.4143  0.5418
#>  -0.4432  0.4533
#>   0.2139 -0.3374
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4146  0.3256  0.4410  0.5379  0.0878  0.0924  0.5187  0.2680 -0.5902
#>   0.1226  0.2133  0.2764  0.6660  0.1768  0.5144  0.3265  0.2039 -0.4193
#>   0.4709  0.4793  0.3231  0.5888 -0.2257  0.4099  0.0385  0.4213  0.0326
#> 
#> Columns 10 to 18 -0.3078 -0.1763 -0.4603  0.4168  0.1521  0.2659  0.2038 -0.4614  0.1558
#>  -0.5721 -0.5399 -0.5057  0.0872  0.4876 -0.3598 -0.2393 -0.2198  0.5444
#>  -0.3405 -0.3940 -0.4142  0.3461  0.2625 -0.3638 -0.0591 -0.5737  0.1720
#> 
#> Columns 19 to 20  0.7953 -0.1105
#>   0.6830  0.0563
#>   0.6386 -0.0738
#> [ CPUFloatType{2,3,20} ]
#>