RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5105 -0.4012  0.4215  0.4093  0.3721  0.0127  0.9369 -0.5556 -0.1557
#>  -0.3636  0.0438 -0.8938 -0.6556 -0.3379  0.4833 -0.1853 -0.9427 -0.7258
#>   0.6258  0.0305 -0.8586 -0.0264  0.6310  0.1213 -0.4194 -0.2313 -0.2260
#> 
#> Columns 10 to 18 -0.8230 -0.3485 -0.3690  0.9298  0.1961 -0.9049 -0.1679 -0.7854 -0.9003
#>  -0.1190  0.5190 -0.5449 -0.2636 -0.5806  0.8280 -0.1279 -0.1768  0.5076
#>  -0.3650  0.4451 -0.2506  0.0811  0.8463 -0.1204 -0.5293  0.2524 -0.3336
#> 
#> Columns 19 to 20  0.0491  0.6762
#>   0.1584 -0.4018
#>   0.6663 -0.0956
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1949 -0.8358  0.3533 -0.2619 -0.5738 -0.6803  0.5361 -0.0569  0.3340
#>  -0.5890  0.0092  0.5310 -0.2488  0.4276  0.4958  0.1605 -0.2544 -0.4709
#>  -0.6242  0.1401 -0.1382 -0.5142  0.5871 -0.0154 -0.1745  0.1575  0.2762
#> 
#> Columns 10 to 18 -0.2563  0.4992  0.1460 -0.1650 -0.5229 -0.5704 -0.0667 -0.4271 -0.0786
#>  -0.3249 -0.2839 -0.1057  0.3066  0.4801 -0.0423 -0.1011 -0.7054 -0.3454
#>  -0.0626  0.1580  0.2911 -0.8295 -0.4124  0.1120  0.1026 -0.6345  0.3105
#> 
#> Columns 19 to 20  0.6760  0.1597
#>  -0.3181  0.5275
#>  -0.2132 -0.0210
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.6610 -0.7430 -0.5134 -0.3135 -0.5682  0.3069  0.0427 -0.6697 -0.4980
#>  -0.3695 -0.5948 -0.2955 -0.4905  0.2149 -0.4336 -0.0671  0.3965  0.3338
#>  -0.6637 -0.4221 -0.0707  0.1345  0.2129  0.3506 -0.1069  0.1019 -0.5987
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1856 -0.1055 -0.4284  0.0594 -0.6264  0.3829  0.8302  0.0966  0.3289
#>  -0.2637  0.4403  0.2527  0.1040  0.0006 -0.0086  0.6091 -0.3213  0.1837
#>   0.7236  0.0575 -0.0639 -0.5136  0.2800 -0.7276 -0.0405  0.1568 -0.4553
#> 
#> Columns 10 to 18 -0.6997 -0.1474  0.5162 -0.8503 -0.5935 -0.6104  0.7466  0.4398  0.0740
#>  -0.5258  0.0625  0.7092 -0.6900 -0.2288  0.3786 -0.2032  0.1685 -0.6408
#>  -0.7006  0.0956  0.0948 -0.6683 -0.3573 -0.1793  0.3599 -0.3442  0.6995
#> 
#> Columns 19 to 20 -0.7192  0.0782
#>  -0.8390 -0.2492
#>  -0.7541 -0.2239
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2012 -0.6706  0.0839 -0.5621  0.1801 -0.4154  0.4995 -0.4433  0.2395
#>  -0.0354 -0.7945  0.3112 -0.2077  0.2731 -0.5067  0.1210 -0.0939  0.0771
#>  -0.0182 -0.5860 -0.1551 -0.5000  0.4792 -0.2038 -0.4046  0.2509 -0.1237
#> 
#> Columns 10 to 18 -0.7389  0.2957  0.0245  0.4296 -0.0392 -0.5297 -0.2379 -0.2170 -0.6336
#>  -0.2183  0.0781 -0.1062  0.4731  0.0671 -0.5199 -0.7465 -0.2900 -0.1928
#>  -0.3571  0.2418  0.0324 -0.1107  0.2188 -0.1244 -0.3291 -0.6257 -0.6986
#> 
#> Columns 19 to 20  0.8030  0.3207
#>   0.1155  0.0577
#>   0.0977 -0.2140
#> [ CPUFloatType{2,3,20} ]
#>