RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8534 -0.8019  0.7934 -0.0897  0.4208 -0.1888 -0.0471 -0.5646  0.2524
#>   0.4741  0.0689  0.0278  0.8094  0.7783  0.3782 -0.6172  0.7605 -0.7489
#>  -0.0708 -0.2665  0.4197  0.8486 -0.0044  0.1366 -0.6257  0.3631 -0.2567
#> 
#> Columns 10 to 18 -0.0212 -0.6811 -0.6510 -0.2476  0.0543 -0.5613  0.0024  0.5926 -0.3180
#>  -0.3733 -0.0749 -0.6171 -0.5201 -0.1496 -0.7062 -0.6213 -0.4765 -0.4310
#>  -0.8018  0.4501 -0.8423  0.1986 -0.2200 -0.6334 -0.2911  0.9788 -0.6979
#> 
#> Columns 19 to 20 -0.6909 -0.7187
#>  -0.1446 -0.1205
#>  -0.1685 -0.4557
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0136  0.1731 -0.3535 -0.7640  0.1358 -0.0881 -0.5017 -0.4454  0.4524
#>  -0.3783 -0.6046  0.1072 -0.2406  0.1936 -0.2736 -0.4523 -0.1818  0.6075
#>   0.2667  0.3088 -0.5699 -0.7029  0.8253 -0.2316 -0.5331  0.1759  0.0952
#> 
#> Columns 10 to 18  0.0069  0.6550 -0.3473  0.0178 -0.0964 -0.3656  0.0718 -0.1021 -0.6035
#>  -0.5154  0.0687 -0.5985  0.0711  0.3848 -0.4039  0.2004  0.6212 -0.3869
#>  -0.1359  0.4413 -0.6425 -0.0902  0.2602 -0.4339 -0.3124 -0.5040 -0.4486
#> 
#> Columns 19 to 20  0.5017  0.2309
#>   0.1702 -0.5962
#>   0.8377 -0.0054
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0204 -0.1887  0.0627 -0.1215 -0.1207  0.2014 -0.3527  0.0995  0.3371
#>   0.4259 -0.3873 -0.5475 -0.1936  0.5417 -0.1931 -0.2181 -0.4032 -0.0104
#>   0.2384 -0.3525  0.1913 -0.4593  0.1280  0.0316  0.0589 -0.0475  0.3381
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3697  0.5613  0.3744 -0.2463  0.0845 -0.3117  0.5028 -0.2234  0.2830
#>   0.8938  0.4569  0.2939 -0.2142  0.3955 -0.4005 -0.3015  0.2540  0.9365
#>  -0.3873  0.3070 -0.7300  0.2113  0.2908 -0.1697  0.6487  0.0443  0.5842
#> 
#> Columns 10 to 18 -0.1285  0.3835  0.5118  0.2650 -0.1772  0.5959  0.1566 -0.1440 -0.6558
#>  -0.6671  0.2739  0.9531  0.6519 -0.1709  0.4898  0.5758  0.0143 -0.8604
#>   0.0123 -0.3437  0.4307  0.0408 -0.3637 -0.6650 -0.4651 -0.0098  0.2693
#> 
#> Columns 19 to 20  0.3979 -0.0565
#>  -0.4936  0.0043
#>   0.8517 -0.1383
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1787 -0.5521 -0.1591 -0.4999 -0.1196 -0.1900 -0.1552  0.2168  0.1499
#>   0.3728  0.0895 -0.4635 -0.7976  0.0607  0.0009 -0.2114  0.1046 -0.0952
#>   0.1586 -0.3280  0.2779 -0.3289  0.4323 -0.2316 -0.2470  0.0402 -0.3807
#> 
#> Columns 10 to 18 -0.4671 -0.3127 -0.5583 -0.1801 -0.3000  0.0777 -0.2164  0.1885 -0.4602
#>  -0.5573 -0.1476 -0.4568 -0.4121 -0.2519  0.2410  0.0445 -0.2200 -0.3266
#>  -0.1798  0.4387 -0.1902  0.2848 -0.2508 -0.6199 -0.2629  0.4351 -0.4641
#> 
#> Columns 19 to 20  0.3915 -0.0752
#>   0.4553  0.1378
#>  -0.0047 -0.0095
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>