RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2979  0.6992  0.5560  0.6459 -0.1916  0.1112 -0.5181  0.2423  0.3035
#>   0.0820  0.6441  0.2188  0.1712 -0.2296 -0.5401 -0.6069 -0.4513  0.7540
#>  -0.2552 -0.3872 -0.3605  0.6006  0.1578  0.5599  0.2633 -0.2739  0.0300
#> 
#> Columns 10 to 18  0.2759 -0.7244 -0.1796 -0.1843 -0.4334 -0.7304 -0.5065  0.4849 -0.4284
#>  -0.2067  0.0624  0.5240  0.4339 -0.2624  0.7426 -0.4038 -0.5282  0.1412
#>   0.5061 -0.5694 -0.3965  0.0075 -0.8141 -0.5593 -0.3044  0.6249 -0.4702
#> 
#> Columns 19 to 20 -0.1876  0.0415
#>  -0.5315 -0.8402
#>  -0.2794  0.7336
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4089  0.2335  0.2637  0.4754 -0.1845 -0.0778 -0.2708  0.0650  0.1585
#>  -0.6441  0.0425  0.1922  0.7141 -0.7285 -0.0006 -0.1351 -0.3077 -0.5653
#>   0.3910  0.7137  0.1662 -0.2499  0.4744  0.0181 -0.6301  0.5637  0.0332
#> 
#> Columns 10 to 18  0.2743  0.3078 -0.1809 -0.4811  0.3472 -0.0731  0.1646 -0.1957  0.0571
#>   0.3368 -0.3967 -0.3370 -0.2157  0.4724 -0.3585 -0.1330  0.2045 -0.0346
#>   0.5820  0.3837  0.3857 -0.2500  0.3544  0.1018  0.1723 -0.0687  0.5093
#> 
#> Columns 19 to 20  0.0592  0.0912
#>  -0.2213  0.2605
#>   0.0094 -0.0719
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0123  0.4044  0.2988  0.3586 -0.3593 -0.2312 -0.2166  0.0055 -0.3845
#>   0.4177  0.4480  0.4364  0.3077 -0.4967  0.0412 -0.6319 -0.0348 -0.2948
#>  -0.6144 -0.3556 -0.0209  0.6682 -0.3851  0.1629  0.4638 -0.2468 -0.1315
#> 
#> Columns 10 to 18  0.8261 -0.0522 -0.0763 -0.3457  0.3518  0.2882 -0.4366  0.1524  0.3367
#>   0.8015  0.4067  0.4343 -0.1025  0.1898  0.3079 -0.4931  0.2081  0.4142
#>  -0.0434 -0.1648 -0.6720 -0.5934  0.2920 -0.4623 -0.3965 -0.0938  0.2249
#> 
#> Columns 19 to 20  0.1914  0.3429
#>   0.5955  0.0103
#>  -0.4888  0.1962
#> 
#> (4,.,.) = 
#>  Columns 1 to 9 -0.4265  0.0074  0.1395  0.7464 -0.6650  0.0566 -0.4900 -0.3001 -0.6601
#>  -0.5250 -0.2035  0.1327  0.2049 -0.6978  0.1358 -0.4380 -0.4091 -0.3722
#>   0.2996  0.5218  0.6125 -0.1352 -0.7057  0.0227 -0.3429 -0.1317  0.5666
#> 
#> Columns 10 to 18  0.5618  0.1633 -0.3455 -0.2708  0.5502  0.1570 -0.5381  0.2351  0.2338
#>   0.4852 -0.4386 -0.5412 -0.2026  0.4605  0.2256 -0.2488  0.6795  0.2209
#>   0.8201 -0.2029  0.1983 -0.2924 -0.3544 -0.2541 -0.2772  0.6571  0.5312
#> 
#> Columns 19 to 20  0.1880 -0.0781
#>  -0.1348 -0.0190
#>   0.2822 -0.2307
#> 
#> (5,.,.) = 
#>  Columns 1 to 9 -0.4630  0.0911  0.4542  0.3191 -0.5756 -0.1458 -0.5609 -0.4550 -0.2115
#>   0.4654  0.2442  0.2294 -0.1651 -0.3368 -0.0991 -0.7361 -0.0422 -0.1908
#>  -0.2292 -0.3728  0.0678  0.6378 -0.2121 -0.4785 -0.4182  0.3822 -0.3264
#> 
#> Columns 10 to 18  0.8266  0.0509 -0.0253 -0.4027  0.1403  0.4931 -0.4582  0.4716  0.1810
#>   0.7194  0.3030  0.0523 -0.2568  0.0281  0.4519 -0.0219  0.4015  0.5689
#>   0.3391  0.0577 -0.6444 -0.3184  0.0442  0.0677 -0.5021 -0.0085  0.4753
#> 
#> Columns 19 to 20  0.1943  0.3411
#>   0.1602 -0.0493
#>   0.3236  0.1083
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.2665 -0.3485 -0.6598 -0.6376 -0.5496 -0.1162  0.5190 -0.0429 -0.5156
#>   0.0571 -0.0814 -0.1611  0.5695 -0.3699  0.8397 -0.4131 -0.1595 -0.5631
#>  -0.0860 -0.7819 -0.7245 -0.3223  0.1671 -0.4728 -0.8076 -0.2861  0.2325
#> 
#> Columns 10 to 18 -0.6107  0.3734 -0.7717 -0.1414  0.4828  0.1894  0.4356  0.8260 -0.8043
#>  -0.1872 -0.1998  0.3901 -0.2471  0.2038  0.4228  0.2101 -0.7067  0.4489
#>   0.0161 -0.4100  0.3023 -0.3581 -0.6496  0.3453 -0.5288 -0.4488  0.2878
#> 
#> Columns 19 to 20  0.8057 -0.5443
#>   0.0554  0.2818
#>  -0.0690 -0.4196
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4630  0.0911  0.4542  0.3191 -0.5756 -0.1458 -0.5609 -0.4550 -0.2115
#>   0.4654  0.2442  0.2294 -0.1651 -0.3368 -0.0991 -0.7361 -0.0422 -0.1908
#>  -0.2292 -0.3728  0.0678  0.6378 -0.2121 -0.4785 -0.4182  0.3822 -0.3264
#> 
#> Columns 10 to 18  0.8266  0.0509 -0.0253 -0.4027  0.1403  0.4931 -0.4582  0.4716  0.1810
#>   0.7194  0.3030  0.0523 -0.2568  0.0281  0.4519 -0.0219  0.4015  0.5689
#>   0.3391  0.0577 -0.6444 -0.3184  0.0442  0.0677 -0.5021 -0.0085  0.4753
#> 
#> Columns 19 to 20  0.1943  0.3411
#>   0.1602 -0.0493
#>   0.3236  0.1083
#> [ CPUFloatType{2,3,20} ]
#>