RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1997  0.5258 -0.4389  0.2175 -0.8074  0.7460 -0.4657  0.8357  0.5188
#>   0.3850  0.3456 -0.4352 -0.8519 -0.4336  0.3986  0.7499  0.3720  0.6593
#>   0.5734  0.5148 -0.0770 -0.8356  0.7257  0.5268  0.7794  0.6317  0.1514
#> 
#> Columns 10 to 18  0.5639  0.4416  0.3924  0.2094 -0.7743  0.8633  0.1196  0.1420  0.6338
#>  -0.7775 -0.6844 -0.4356 -0.4004  0.2408 -0.2209 -0.6650  0.3566 -0.5140
#>  -0.3532  0.4611  0.5171 -0.1275 -0.4146  0.0171 -0.2450  0.4424 -0.3717
#> 
#> Columns 19 to 20 -0.7622 -0.6586
#>  -0.7873 -0.7027
#>  -0.7234  0.1584
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4277 -0.3482 -0.0444 -0.3429  0.3484  0.3067  0.4067  0.3523  0.4958
#>   0.4337  0.3424  0.4154 -0.4487  0.2251 -0.2833  0.4429  0.4592  0.7050
#>  -0.4333  0.4745  0.3919  0.0907 -0.0535 -0.0334  0.6746  0.5175  0.1916
#> 
#> Columns 10 to 18  0.4364  0.5704  0.2531  0.6635 -0.1696 -0.0520 -0.3792  0.5528  0.0593
#>   0.1845  0.2201 -0.1076 -0.0190 -0.0459  0.0229 -0.0818  0.6579  0.1495
#>   0.0836  0.4260  0.2501  0.0599  0.2029 -0.3778 -0.1913  0.1825  0.1152
#> 
#> Columns 19 to 20  0.0673 -0.5608
#>  -0.1475  0.2163
#>  -0.3067 -0.0970
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3762  0.1345  0.5125 -0.2141 -0.2057  0.3851  0.1150  0.3148 -0.1168
#>   0.1662  0.2313  0.2615 -0.1978  0.3028  0.2571  0.7532  0.2518 -0.0566
#>  -0.0126  0.1932  0.3402 -0.2538 -0.1676  0.0625  0.4490  0.1836 -0.1160
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5653  0.6728 -0.8053  0.7717  0.6201  0.5753  0.4389  0.3289 -0.1409
#>   0.0295  0.5618  0.0501  0.3350  0.2352  0.0142  0.7057  0.3413  0.1375
#>   0.5425  0.4903 -0.0361  0.4014 -0.4021 -0.4629  0.7387  0.2092  0.0680
#> 
#> Columns 10 to 18 -0.7506  0.1630 -0.5680 -0.2535  0.7577  0.4041 -0.3851  0.9341  0.3340
#>  -0.7613 -0.1285  0.4104  0.3749  0.3485  0.7286  0.0103  0.5815 -0.1227
#>  -0.4103 -0.1956  0.0281  0.7013  0.6314  0.5838 -0.4773 -0.2329  0.4723
#> 
#> Columns 19 to 20 -0.4829  0.5604
#>  -0.5173  0.4608
#>  -0.2806  0.1060
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0049  0.3510  0.2522 -0.3732 -0.0138  0.3241  0.0230  0.1269 -0.4605
#>   0.1668  0.3150 -0.0771 -0.4360 -0.2848  0.1348  0.0334  0.4099  0.1753
#>   0.2848  0.5279 -0.1150 -0.2528 -0.2243  0.2835  0.1626  0.5739  0.2424
#> 
#> Columns 10 to 18  0.3912  0.8142 -0.1563  0.2185  0.5350  0.0666 -0.0913  0.3195 -0.6362
#>   0.2635  0.5898  0.2004  0.2971 -0.2074 -0.1862  0.0302  0.4431 -0.1940
#>   0.1656  0.2917  0.3436  0.2223 -0.0980 -0.2480 -0.1402  0.3502 -0.2725
#> 
#> Columns 19 to 20 -0.1172 -0.2265
#>  -0.3531 -0.5915
#>  -0.4378 -0.4255
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>