RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2091 -0.6457 -0.0166  0.3106 -0.4947 -0.0182  0.0130  0.3626  0.0696
#>   0.5108  0.0866 -0.6574  0.3095 -0.6135  0.4461  0.5558  0.4022  0.0249
#>   0.9000 -0.4609 -0.3524 -0.9509 -0.9099 -0.1753  0.9774  0.7540  0.6024
#> 
#> Columns 10 to 18  0.4264  0.2722 -0.1961  0.4266 -0.8714 -0.4213  0.9568 -0.6349 -0.7368
#>  -0.2329  0.2175 -0.6330  0.5938 -0.7811 -0.0152  0.8584 -0.4628  0.7113
#>  -0.5003  0.1409  0.3452  0.9676 -0.0827  0.5271  0.6599  0.6174 -0.1340
#> 
#> Columns 19 to 20  0.4721  0.7200
#>   0.3676 -0.3903
#>  -0.7928 -0.2839
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0723 -0.4810 -0.5926  0.0892  0.5696 -0.0518  0.2880  0.0716  0.4025
#>  -0.2176 -0.4436 -0.6369  0.4069  0.5811 -0.4649  0.0626  0.2469  0.4771
#>   0.7340 -0.6354 -0.3219 -0.0375  0.4625 -0.1715  0.0784 -0.2837 -0.2179
#> 
#> Columns 10 to 18  0.3160  0.4294  0.3146 -0.2487  0.0467 -0.4890  0.2815 -0.0902 -0.5824
#>  -0.1274 -0.0251  0.4630  0.1474  0.0595  0.4301 -0.0640  0.5340 -0.4562
#>  -0.2000 -0.2649  0.1455  0.2515 -0.1421  0.0017 -0.3898 -0.4242 -0.2669
#> 
#> Columns 19 to 20 -0.5730  0.2165
#>  -0.1231 -0.1215
#>  -0.0763 -0.3573
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3490 -0.3265 -0.5440  0.5662 -0.2112  0.1148 -0.4645 -0.0477  0.1521
#>   0.4891 -0.3683 -0.5131  0.0398  0.2793 -0.2297 -0.5720 -0.0099  0.0697
#>  -0.2953 -0.5663 -0.6319 -0.2871 -0.5264 -0.3179 -0.1072  0.4149 -0.2672
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1523 -0.1954  0.4107  0.4370 -0.5916  0.3883  0.0380 -0.3557  0.1780
#>   0.0664 -0.8076  0.2327  0.6792 -0.7111  0.0642 -0.2723  0.0643  0.1160
#>   0.4000 -0.5008 -0.4274 -0.1929 -0.1166 -0.6216 -0.2845 -0.5326 -0.0415
#> 
#> Columns 10 to 18 -0.2870 -0.3341  0.5552  0.4475 -0.7844  0.5432 -0.4832  0.2644  0.1071
#>  -0.1099  0.2971 -0.0702  0.7700 -0.7551 -0.6191 -0.5748  0.4193 -0.2196
#>  -0.0800 -0.2605 -0.4320 -0.3963 -0.3105 -0.5042 -0.3212 -0.4855  0.2414
#> 
#> Columns 19 to 20  0.5469  0.6851
#>  -0.4063  0.5736
#>   0.5122 -0.1484
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2261 -0.3617 -0.4038  0.2093  0.2425  0.3064  0.1149  0.0794  0.4305
#>  -0.0724 -0.6828 -0.4320  0.2830  0.1304 -0.2041 -0.3694  0.2058  0.1532
#>   0.0431 -0.2364 -0.6840  0.3576  0.5792  0.0745 -0.4306 -0.2005  0.0118
#> 
#> Columns 10 to 18  0.0930  0.3144  0.0936  0.0698  0.0085 -0.3529  0.4906  0.1889 -0.2533
#>  -0.1064  0.3872  0.0714  0.0158 -0.3595 -0.1363  0.7385  0.0661 -0.3856
#>  -0.0802  0.0301 -0.1772 -0.2166  0.2457  0.2247 -0.3601 -0.2845 -0.2074
#> 
#> Columns 19 to 20 -0.4731  0.2401
#>  -0.5317  0.2010
#>  -0.2049 -0.1120
#> [ CPUFloatType{2,3,20} ]
#>