RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0645 -0.7151  0.4030  0.6194  0.0915  0.5064  0.3157  0.6380 -0.5100
#>   0.9283 -0.1626 -0.5848 -0.7107  0.4909  0.5882  0.1729 -0.8373 -0.3387
#>   0.2893 -0.4949 -0.0813  0.2460  0.0777 -0.0019 -0.3771  0.1285 -0.7091
#> 
#> Columns 10 to 18  0.2855  0.4393  0.9143 -0.2401 -0.1371 -0.1460  0.4198  0.2337  0.0629
#>  -0.3594 -0.5958  0.6240 -0.2087  0.7153 -0.9017  0.1814 -0.9336  0.4500
#>   0.1616  0.7846  0.6330  0.2923 -0.6049  0.3877  0.4887 -0.6222 -0.0977
#> 
#> Columns 19 to 20  0.8695  0.4455
#>   0.6522 -0.4302
#>   0.3895  0.8030
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0322 -0.1474  0.2818 -0.2011 -0.3336 -0.3013  0.0152  0.3424  0.3235
#>   0.1530  0.0466  0.1854  0.4506 -0.4394  0.6117  0.2305  0.0331  0.5230
#>  -0.2592 -0.6418 -0.1237 -0.2023 -0.4307 -0.6949 -0.2865 -0.2208  0.2030
#> 
#> Columns 10 to 18 -0.2896  0.5048  0.5604 -0.4432 -0.0979 -0.1677 -0.0923  0.3244  0.1810
#>  -0.2771  0.1286  0.7377  0.3862 -0.4774 -0.5511  0.7173  0.5162 -0.4819
#>  -0.1864 -0.0273  0.1936 -0.0522 -0.1734  0.3942  0.2428  0.2689 -0.0626
#> 
#> Columns 19 to 20  0.3296 -0.1165
#>   0.2238 -0.6493
#>   0.3320 -0.2834
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2491 -0.5201 -0.4435 -0.1528 -0.1880 -0.4311 -0.6631 -0.5299 -0.2058
#>   0.1284  0.3882 -0.5061 -0.3593 -0.3673  0.0143  0.2590  0.3225 -0.6434
#>  -0.1212 -0.1831  0.3092  0.1289 -0.3480 -0.0090 -0.6691  0.2800  0.2506
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5404  0.2199  0.1002 -0.5134  0.3451 -0.6079 -0.0957 -0.7698 -0.2415
#>   0.2210 -0.6158  0.6117  0.4439 -0.5203 -0.0642  0.2349 -0.4445 -0.1169
#>  -0.0468  0.3406  0.2328 -0.3855 -0.2573 -0.2552  0.5526 -0.3695 -0.7344
#> 
#> Columns 10 to 18 -0.2424  0.0934 -0.7757 -0.7518  0.5355  0.2621 -0.6985 -0.5014 -0.5362
#>   0.6731  0.3676 -0.1110  0.6677 -0.0006 -0.2392  0.6195 -0.2572  0.5508
#>  -0.7019  0.2611  0.4594 -0.8639  0.3261  0.2458 -0.5368 -0.4359 -0.4634
#> 
#> Columns 19 to 20 -0.5484  0.3995
#>   0.4056  0.3177
#>  -0.3064  0.1549
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3409  0.1864 -0.2694 -0.4454 -0.4857 -0.2018 -0.4308 -0.5307  0.1576
#>   0.2492 -0.5275  0.2001  0.4636 -0.1903  0.0997 -0.7215  0.0829  0.0706
#>  -0.1208 -0.1487 -0.3836 -0.4806 -0.3687 -0.0832 -0.3536 -0.3442  0.1805
#> 
#> Columns 10 to 18 -0.7031  0.2194  0.4141 -0.1534  0.0128 -0.3616 -0.2647  0.3279 -0.1788
#>  -0.1907  0.4896  0.4797 -0.3771  0.0588 -0.6048  0.3284 -0.1789 -0.6923
#>  -0.5418  0.5463  0.4147 -0.3555 -0.2966  0.1432 -0.0953  0.0526 -0.1655
#> 
#> Columns 19 to 20 -0.3200  0.4003
#>   0.4447 -0.3009
#>   0.0826  0.0790
#> [ CPUFloatType{2,3,20} ]
#>