RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1451  0.4029 -0.3840  0.3589 -0.5842 -0.7876  0.1915 -0.3902 -0.6513
#>  -0.1525 -0.0899 -0.7186  0.1242 -0.0103  0.7619  0.3659 -0.5866 -0.3088
#>  -0.4530 -0.2335  0.7216  0.7954 -0.8534  0.7181 -0.6812 -0.8553  0.4851
#> 
#> Columns 10 to 18  0.4632  0.2859 -0.2077 -0.8368  0.3631 -0.0842  0.6548 -0.5706  0.5899
#>   0.5010 -0.0351  0.7063 -0.5801  0.6337  0.4692  0.7633 -0.9280  0.9300
#>  -0.0893  0.7043 -0.6453  0.0293  0.0743 -0.2040  0.3199 -0.6397 -0.7496
#> 
#> Columns 19 to 20 -0.3618 -0.6475
#>  -0.0682 -0.2184
#>   0.7826 -0.5293
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0471  0.0689  0.2849  0.3914 -0.6891  0.3717  0.5767  0.4240  0.2377
#>   0.0420  0.2873  0.2350 -0.1232 -0.2311  0.3509 -0.0726  0.5432  0.3378
#>   0.3359 -0.1224  0.1413 -0.4989 -0.4160  0.4475 -0.0084 -0.2401  0.4040
#> 
#> Columns 10 to 18 -0.2924 -0.0703 -0.2724  0.6353 -0.3019  0.0107 -0.5006 -0.1825 -0.2560
#>   0.2499  0.3177  0.2295  0.8195  0.0937 -0.3083  0.3874 -0.1423 -0.3621
#>  -0.7536  0.2103  0.4112  0.3871  0.4195  0.1900 -0.3954 -0.3375 -0.2992
#> 
#> Columns 19 to 20 -0.6522  0.3099
#>   0.1541 -0.3204
#>  -0.0404 -0.8743
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0260 -0.1585 -0.3142  0.4403 -0.5546 -0.0644 -0.1237 -0.0409 -0.6198
#>   0.3309 -0.3900  0.2040 -0.1487  0.1639 -0.2546 -0.2463 -0.0423  0.0595
#>   0.4912  0.1081  0.5043  0.2212 -0.1154 -0.4691 -0.0763  0.3474 -0.3991
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0674 -0.0769 -0.2750  0.6723 -0.4740 -0.1836 -0.6125  0.0141 -0.1011
#>  -0.2790  0.0904  0.3658  0.8093  0.3827 -0.4769 -0.0704 -0.7966  0.2712
#>   0.1589 -0.4760 -0.1783 -0.0028 -0.6422  0.3330 -0.2519  0.4025 -0.6264
#> 
#> Columns 10 to 18  0.6656  0.4671  0.8539 -0.3833 -0.4617 -0.2309 -0.4814  0.2577  0.4401
#>   0.3021  0.1525  0.3505 -0.0762 -0.8025  0.8247 -0.4629  0.1079  0.2229
#>   0.8033 -0.1907  0.4871  0.4419 -0.6080 -0.2354 -0.5989 -0.4872  0.5597
#> 
#> Columns 19 to 20  0.1190  0.7076
#>  -0.0945  0.3596
#>  -0.5652  0.7253
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2499 -0.0424 -0.0580  0.2505 -0.5680  0.4067  0.2561  0.4200  0.0242
#>   0.2535 -0.1295  0.2626 -0.4351 -0.4111  0.2496  0.3346 -0.0697  0.3309
#>  -0.0434  0.1137 -0.1842  0.2576 -0.8439  0.0773  0.1783  0.0748 -0.4321
#> 
#> Columns 10 to 18  0.0366  0.8422  0.0596  0.5138  0.1243  0.0252  0.4386 -0.6480  0.0057
#>  -0.6033  0.6638 -0.4118  0.0128  0.0674 -0.4633 -0.1170 -0.4717 -0.3419
#>  -0.2307  0.5857 -0.3724  0.1604  0.4878  0.2763  0.3912 -0.1551 -0.3584
#> 
#> Columns 19 to 20 -0.2549 -0.4441
#>  -0.4527 -0.2416
#>  -0.4951 -0.5851
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>