RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4341 -0.7481 -0.0155 -0.1667 -0.1737  0.9354 -0.1024  0.6209  0.7331
#>   0.4818  0.6665 -0.7605 -0.0685  0.0970 -0.2761 -0.6364 -0.5194  0.5688
#>  -0.2410  0.8041 -0.5350 -0.1154  0.0720 -0.1012 -0.3394 -0.8164 -0.5111
#> 
#> Columns 10 to 18 -0.6326  0.4706  0.6197  0.8631  0.3008  0.0961  0.3755 -0.8333 -0.2493
#>  -0.6738 -0.5237  0.2451 -0.4069 -0.5264  0.6043 -0.4336 -0.7208  0.7170
#>   0.4165 -0.8329 -0.0629 -0.1459  0.3747  0.0225 -0.0177 -0.3828  0.2856
#> 
#> Columns 19 to 20  0.8207 -0.2897
#>  -0.0639 -0.2977
#>   0.6821  0.3876
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1063  0.3392 -0.1736 -0.0132 -0.2353  0.3109 -0.1159 -0.7230 -0.2804
#>   0.3689  0.4177  0.1049  0.3026  0.2117  0.1271 -0.3236 -0.4954  0.1402
#>  -0.2533 -0.4031  0.2856  0.1993  0.4932  0.1930 -0.0828 -0.3276  0.3202
#> 
#> Columns 10 to 18 -0.3410 -0.2164  0.6276  0.3368 -0.0167  0.3719 -0.7188 -0.6864 -0.2286
#>   0.3881 -0.0402  0.1149 -0.3783 -0.1443  0.4664 -0.6964 -0.8231  0.5851
#>   0.1323 -0.1871  0.0928 -0.1782  0.2222 -0.1331 -0.1200 -0.4948  0.4937
#> 
#> Columns 19 to 20  0.7209 -0.1616
#>   0.5246  0.4013
#>   0.4094  0.2186
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0737  0.2822 -0.0505 -0.1154 -0.2111  0.3229  0.3124 -0.2327 -0.1863
#>   0.4487 -0.4510 -0.2076 -0.3074  0.1834  0.0303 -0.4213 -0.1905  0.0647
#>   0.3280  0.2640 -0.0669 -0.0208 -0.0807 -0.1856 -0.2303 -0.2598  0.0174
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0115  0.0240 -0.2924  0.8755 -0.1208 -0.4961 -0.4627 -0.5664  0.3328
#>   0.6423 -0.7176  0.1257  0.6481 -0.0629  0.1632  0.7915 -0.1701 -0.3607
#>   0.6125 -0.1774  0.6561  0.3837 -0.0462 -0.0686  0.8081 -0.6459 -0.6129
#> 
#> Columns 10 to 18  0.0217  0.3140  0.4306  0.1964  0.3051 -0.1385  0.1522 -0.5736  0.3529
#>   0.0952  0.0326 -0.3462  0.1499 -0.3696 -0.2271 -0.2429 -0.1532 -0.0242
#>   0.4828  0.7583  0.8195  0.0678  0.5375  0.8486 -0.5550  0.6793 -0.3351
#> 
#> Columns 19 to 20  0.5596 -0.0410
#>  -0.5400 -0.4049
#>  -0.3541 -0.0970
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0298 -0.0463 -0.3046 -0.1967  0.4208  0.1621  0.0368 -0.2495  0.3218
#>   0.6373  0.4638  0.2825 -0.0159 -0.3832 -0.1974 -0.4051 -0.1644 -0.2495
#>   0.4224  0.5936  0.4294 -0.0868 -0.0166 -0.0015 -0.3633 -0.7045 -0.0894
#> 
#> Columns 10 to 18 -0.1435 -0.2809 -0.1815 -0.5424  0.0836  0.3839 -0.3797 -0.6542  0.6884
#>   0.1865 -0.0996  0.0635  0.0786  0.0125 -0.0992 -0.2241 -0.2546  0.7308
#>   0.1284 -0.4054  0.1127 -0.4355 -0.1516 -0.3504 -0.4453 -0.5662  0.6315
#> 
#> Columns 19 to 20  0.0415 -0.2396
#>   0.3656  0.0484
#>   0.4587  0.2440
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>