RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1488 -0.6411 -0.7568 -0.7953 -0.1327  0.1594  0.2826  0.7375 -0.4185
#>  -0.7600  0.3211 -0.7543 -0.9127  0.6047 -0.2737  0.3788  0.0431  0.1452
#>   0.7319  0.0671  0.7470  0.0715 -0.1361 -0.9200 -0.9395  0.5499 -0.3458
#> 
#> Columns 10 to 18  0.3798 -0.6422  0.0315 -0.0822  0.3724  0.3802  0.2083 -0.6428  0.7253
#>  -0.8442 -0.8222 -0.7041 -0.2945  0.3655 -0.4035  0.5971 -0.7242 -0.0801
#>   0.1289 -0.6868  0.6791  0.3313  0.7345  0.4904  0.7351 -0.4065  0.8224
#> 
#> Columns 19 to 20  0.6728  0.4461
#>   0.1497 -0.2755
#>   0.5931  0.7099
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3201 -0.3683  0.5014 -0.0867 -0.3351  0.4654  0.1687 -0.6065  0.5582
#>   0.4458 -0.8089  0.0964  0.0941 -0.4123  0.1868  0.5379 -0.3903  0.2653
#>   0.6243 -0.0224  0.5624 -0.2194 -0.2129 -0.0572  0.5026 -0.8352 -0.2566
#> 
#> Columns 10 to 18 -0.6373 -0.6408 -0.3285 -0.2800  0.4087  0.1368 -0.0331 -0.2056  0.0015
#>  -0.4051 -0.3941  0.0608 -0.1275  0.2875  0.3031 -0.6622 -0.3036 -0.2301
#>  -0.4885 -0.7043  0.1899  0.1131 -0.2623  0.0688 -0.7059 -0.2764  0.3264
#> 
#> Columns 19 to 20 -0.3125  0.0068
#>  -0.0387 -0.5194
#>  -0.7934 -0.6469
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4879 -0.4153 -0.2140 -0.2849  0.2394  0.6095 -0.1024 -0.1822 -0.3930
#>   0.4951 -0.6814 -0.1601 -0.2955  0.3579  0.6207  0.0778 -0.4526 -0.4729
#>  -0.5138 -0.2504 -0.3618 -0.4388  0.5681  0.3571 -0.2806 -0.5131 -0.3610
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1732 -0.5359  0.4771  0.0673  0.0285  0.3784  0.5879  0.7487 -0.6151
#>  -0.2195  0.0751 -0.4235  0.4077 -0.2574  0.5078 -0.5713  0.1581 -0.6433
#>   0.0845 -0.4555  0.0062  0.6824  0.4098 -0.5075  0.2632 -0.2893 -0.7053
#> 
#> Columns 10 to 18 -0.4051 -0.5604  0.1541  0.1340  0.3665  0.2512  0.4928  0.8613  0.5146
#>   0.2928  0.7129 -0.4017  0.2628  0.5984  0.3434  0.5884 -0.1829 -0.0288
#>  -0.4807 -0.5463  0.3531 -0.2743 -0.0907 -0.4364 -0.3053  0.5415 -0.3208
#> 
#> Columns 19 to 20  0.5697  0.9421
#>   0.4031  0.5246
#>  -0.5776  0.3286
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0059 -0.6491 -0.1614 -0.8643  0.1529  0.1279  0.1236 -0.7224  0.5453
#>  -0.0615 -0.0971 -0.2996 -0.7099  0.1264  0.0131  0.1500 -0.6109 -0.0737
#>   0.3634 -0.6496 -0.1086 -0.7306 -0.2188  0.2873  0.5337 -0.1492 -0.5534
#> 
#> Columns 10 to 18 -0.8886 -0.5169 -0.4720 -0.3898 -0.5183 -0.2154  0.1396  0.3626  0.1445
#>  -0.6852 -0.4732 -0.0735 -0.1253 -0.1696  0.3223 -0.1545  0.1262  0.4414
#>  -0.3466 -0.3548  0.0587  0.1345 -0.3021 -0.1905  0.0827  0.3139  0.5032
#> 
#> Columns 19 to 20  0.0068  0.2291
#>   0.0172  0.4015
#>  -0.1244  0.0104
#> [ CPUFloatType{2,3,20} ]
#>