RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1220 -0.4387  0.1952  0.1895  0.9116  0.5358  0.5050  0.6432 -0.3694
#>  -0.6356 -0.6938 -0.7456 -0.6122 -0.9380  0.7042  0.6735 -0.9015 -0.2675
#>   0.7736  0.6820 -0.4765  0.7239  0.7906  0.4898 -0.1291  0.6213  0.6189
#> 
#> Columns 10 to 18 -0.0348 -0.8010  0.3674  0.6247 -0.2819 -0.0501  0.5592  0.5194 -0.8053
#>  -0.6346 -0.8917  0.1790  0.0681 -0.8403  0.3213 -0.4852  0.3654  0.2507
#>   0.0045  0.7212  0.0809  0.0669  0.2702 -0.8271  0.5698 -0.1984 -0.1057
#> 
#> Columns 19 to 20 -0.4820  0.5245
#>   0.3079  0.1111
#>   0.4001 -0.4512
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1442 -0.4326  0.2377 -0.5712  0.2038  0.0710  0.5160 -0.4780 -0.2960
#>  -0.2216  0.0049 -0.2429 -0.4344  0.0124  0.2563 -0.4427  0.0108 -0.6779
#>  -0.2903 -0.3300 -0.5585 -0.1120  0.0851  0.0915  0.6503 -0.1847  0.4071
#> 
#> Columns 10 to 18 -0.0313 -0.4022 -0.1796  0.3509 -0.3313  0.1032 -0.2565 -0.5187  0.2426
#>   0.3720  0.4184 -0.4782  0.7338 -0.3073  0.1287 -0.1428  0.0228  0.4330
#>  -0.6588 -0.6387 -0.3078 -0.0434 -0.3277  0.2171 -0.2155 -0.6307 -0.5991
#> 
#> Columns 19 to 20  0.3557  0.3254
#>  -0.0221  0.1645
#>   0.1279 -0.2343
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2272  0.3507 -0.1790  0.2154  0.1658 -0.4779  0.0640 -0.1248 -0.7396
#>  -0.4318 -0.3424 -0.1436 -0.1099 -0.2499  0.3658  0.0447 -0.4136 -0.0520
#>  -0.2393 -0.0145 -0.1452 -0.6134  0.1265 -0.0955  0.3277 -0.1688 -0.1789
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4011  0.1989  0.1191 -0.2825 -0.6629  0.5878  0.4473  0.4348 -0.0818
#>  -0.0746  0.6552 -0.6820  0.8512 -0.5817  0.0375 -0.1800  0.8594  0.6691
#>  -0.3390  0.5820  0.2810 -0.2261  0.4043 -0.4384  0.1939  0.5067 -0.1244
#> 
#> Columns 10 to 18 -0.4194 -0.4374 -0.3774 -0.0522 -0.3161  0.1209  0.4486 -0.5049  0.1045
#>  -0.1249 -0.0426  0.1827 -0.7260 -0.1884 -0.5997  0.0370  0.0655 -0.9403
#>  -0.1009 -0.8698  0.0383  0.5212  0.2298 -0.2570 -0.3694  0.3493  0.0404
#> 
#> Columns 19 to 20  0.1714 -0.5723
#>   0.6401  0.7595
#>   0.3906  0.2102
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1878 -0.0036 -0.1005 -0.7092  0.3251  0.4346 -0.1385  0.1722 -0.1027
#>   0.3025  0.3957 -0.2293 -0.2389  0.4140  0.2079  0.3616 -0.1716  0.2259
#>  -0.5511 -0.6036  0.0994 -0.2142  0.0701  0.2578 -0.1116  0.0583  0.1356
#> 
#> Columns 10 to 18 -0.0126 -0.1218 -0.0534  0.3498  0.1736  0.2829 -0.4601 -0.4110 -0.1129
#>  -0.4191 -0.0203  0.3507  0.5366  0.2089  0.0888 -0.0024 -0.1155 -0.5355
#>   0.3815 -0.3206 -0.3314  0.1996 -0.0076  0.4266 -0.4039 -0.2581 -0.1157
#> 
#> Columns 19 to 20  0.0266  0.0618
#>   0.1998 -0.8159
#>  -0.2086 -0.5562
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>