RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0158 -0.0937 -0.8264  0.1438 -0.5916  0.0713  0.6530 -0.6098 -0.0865
#>   0.4862 -0.4408 -0.1927 -0.0267 -0.4707  0.2311 -0.2332 -0.9518 -0.3112
#>  -0.4285  0.6810 -0.6008  0.1563 -0.6821  0.8813  0.3911 -0.0284  0.5856
#> 
#> Columns 10 to 18 -0.8309  0.3964  0.0397 -0.1627  0.2563  0.9594 -0.7215  0.5024 -0.0533
#>   0.6089  0.4739  0.7156  0.4009  0.4723 -0.5448 -0.7906  0.5081  0.6938
#>  -0.6193 -0.2240  0.4182 -0.8359  0.7553  0.5913  0.3247  0.3899  0.4829
#> 
#> Columns 19 to 20  0.2894  0.2357
#>  -0.7647  0.8000
#>   0.9227 -0.6286
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0445  0.2013 -0.4552  0.2612 -0.0619 -0.4122 -0.3544  0.0644  0.2633
#>   0.0495 -0.3225 -0.3073  0.6698 -0.7796  0.4317 -0.7104 -0.7684 -0.0787
#>   0.3987 -0.2619 -0.5146  0.2402 -0.3106 -0.7184  0.1126  0.1416  0.0759
#> 
#> Columns 10 to 18 -0.1834 -0.1532  0.3971 -0.1694  0.0902  0.4433 -0.4972 -0.1688  0.2300
#>   0.7653  0.2746  0.7942 -0.2152 -0.0058 -0.5571 -0.0880  0.3298  0.3485
#>  -0.1563 -0.4531  0.1806 -0.1265  0.0119  0.3316 -0.5868  0.5408 -0.0660
#> 
#> Columns 19 to 20  0.6313  0.3231
#>  -0.2229 -0.0886
#>  -0.0783  0.5818
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2725  0.3726 -0.6119  0.2393 -0.3261  0.0336 -0.1539  0.2676  0.1795
#>   0.0802 -0.2364 -0.7424  0.4162 -0.7487  0.3467 -0.3057 -0.6656 -0.2551
#>  -0.5043  0.2714 -0.6181  0.2431 -0.0965  0.5640 -0.0366 -0.4695  0.0723
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0957  0.2134 -0.2880  0.3482  0.1699 -0.1926 -0.0182  0.3722  0.0231
#>  -0.1133 -0.3144  0.2543  0.0909  0.7429  0.3667  0.6820  0.1390 -0.5151
#>   0.5028  0.0445  0.1161 -0.0208  0.0140 -0.0536  0.1279 -0.1486 -0.0886
#> 
#> Columns 10 to 18 -0.7744  0.5903  0.2582  0.1503 -0.7897  0.0695 -0.1835 -0.1347  0.1960
#>   0.5217 -0.1211 -0.3607  0.6174 -0.1946 -0.5075  0.4662  0.5238 -0.3126
#>   0.0373  0.0060  0.4099  0.0689 -0.6966 -0.1598  0.0485 -0.7441  0.2899
#> 
#> Columns 19 to 20 -0.2505 -0.0449
#>  -0.2996  0.5963
#>  -0.2384  0.0357
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2733  0.2883 -0.4949  0.2315 -0.0842  0.0132 -0.2414 -0.0555  0.0732
#>   0.1454  0.2308 -0.3538 -0.1446 -0.0652 -0.0035  0.4347 -0.3288  0.5094
#>  -0.0749  0.3403 -0.4521  0.2433 -0.0868  0.0901 -0.2954 -0.1087  0.5026
#> 
#> Columns 10 to 18 -0.3642 -0.3413  0.0199  0.0546  0.2481  0.4321 -0.2549  0.2812 -0.0735
#>  -0.4470 -0.0414 -0.3663  0.0341 -0.4779  0.5673 -0.0645  0.1375  0.2989
#>   0.0873  0.3477  0.1937 -0.2956  0.2600  0.1706 -0.4761  0.2451  0.0727
#> 
#> Columns 19 to 20 -0.2044  0.1491
#>  -0.2234  0.3536
#>   0.2630  0.0276
#> [ CPUFloatType{2,3,20} ]
#>