RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1426 -0.3726  0.0014  0.1415 -0.1159  0.2926 -0.4093  0.5326  0.3574
#>   0.6985  0.5315 -0.6121 -0.6003  0.1755  0.7135  0.7094 -0.4819  0.3697
#>   0.6886  0.7783 -0.1861 -0.5913  0.2447  0.7046  0.2029 -0.1785  0.7856
#> 
#> Columns 10 to 18 -0.1292  0.3951  0.5466  0.5628 -0.0904  0.5135  0.5018  0.1117 -0.6775
#>  -0.2881 -0.0881 -0.7602 -0.8854  0.7456  0.4353  0.1395  0.4941  0.0811
#>   0.5490  0.1541 -0.8946 -0.9474 -0.6296  0.6986  0.1740  0.8606 -0.3558
#> 
#> Columns 19 to 20 -0.4566 -0.1164
#>  -0.1981 -0.7841
#>   0.5237 -0.6906
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3014 -0.3091 -0.3982  0.0377 -0.3976  0.2353  0.0708 -0.1211  0.3590
#>   0.0848  0.7421 -0.0638  0.1102  0.0026  0.3907  0.3257 -0.0506  0.5054
#>   0.5378  0.7475 -0.2360 -0.2195  0.0137  0.3076  0.4215  0.3509  0.0853
#> 
#> Columns 10 to 18 -0.1487 -0.1543 -0.1444 -0.3384 -0.5692  0.2759  0.5163  0.1606 -0.0547
#>   0.1843  0.6910 -0.2353 -0.8858 -0.4671  0.5931  0.5866  0.5189  0.2003
#>   0.0143 -0.0486 -0.7711 -0.8966 -0.2552  0.6085  0.3775  0.7106  0.4267
#> 
#> Columns 19 to 20 -0.6025 -0.0061
#>   0.2256 -0.1237
#>   0.1400  0.0799
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0518  0.5793 -0.0785 -0.2449 -0.1683  0.5074  0.4124 -0.0439  0.5963
#>   0.4466  0.1603 -0.2825  0.0339  0.0438  0.1050 -0.2453  0.3370  0.3439
#>   0.7106  0.5536 -0.4388  0.5080  0.0374  0.5392  0.0413 -0.0013  0.2337
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5120  0.6327  0.1875  0.5206 -0.4549  0.3505 -0.5167 -0.7644  0.6510
#>   0.1216 -0.5076  0.4762  0.3789 -0.5533 -0.1587  0.5208 -0.1966  0.5580
#>  -0.7388 -0.6800 -0.6436 -0.7427  0.3398 -0.2122  0.3289 -0.4582 -0.2866
#> 
#> Columns 10 to 18  0.4930  0.2868  0.0684 -0.4048 -0.4235  0.3520  0.3725 -0.7246  0.2075
#>  -0.1392  0.1291  0.2150 -0.7283  0.4782 -0.3855  0.4859  0.5356  0.6081
#>   0.0264 -0.6633 -0.0143  0.6006 -0.1720 -0.8022  0.2192 -0.1098 -0.4734
#> 
#> Columns 19 to 20  0.1976 -0.5727
#>  -0.5713  0.0963
#>  -0.0236 -0.4080
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1670  0.4592 -0.3703  0.0094  0.0660  0.5964  0.5804  0.5244  0.3788
#>   0.3651  0.4912 -0.2296 -0.0028 -0.2974  0.7169  0.5595 -0.3150  0.3396
#>   0.6385  0.0434 -0.3974 -0.3935  0.1562  0.1957 -0.3422 -0.3130  0.2450
#> 
#> Columns 10 to 18  0.1186 -0.2124 -0.2910 -0.4812  0.1224  0.2633  0.2641  0.6594 -0.1947
#>   0.2513  0.2520 -0.6223 -0.7716 -0.0740  0.2826  0.3030  0.4401  0.0802
#>  -0.1599 -0.3475 -0.1000 -0.0290  0.3450  0.1116 -0.2237  0.4390  0.1440
#> 
#> Columns 19 to 20 -0.3771  0.0487
#>  -0.0591  0.1172
#>  -0.0936 -0.4444
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>