RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6330 -0.1159 -0.8227  0.3353  0.3270  0.2044  0.6870  0.8481 -0.0430
#>   0.2029 -0.4450 -0.4610  0.2687  0.8057 -0.1178 -0.6057  0.4468  0.0628
#>  -0.4048  0.0379  0.3248 -0.2097  0.1309  0.0884  0.6341  0.9654  0.5810
#> 
#> Columns 10 to 18 -0.1163 -0.0776  0.4254  0.1297  0.7435 -0.4028  0.5459 -0.4199 -0.3359
#>  -0.0493 -0.1369 -0.5679 -0.1443  0.8855  0.3880  0.6120 -0.9128  0.5618
#>   0.4363 -0.0324  0.0412 -0.4879  0.6358 -0.3463 -0.1071 -0.3148  0.8289
#> 
#> Columns 19 to 20 -0.1804 -0.6574
#>   0.1361  0.1326
#>  -0.9210 -0.8175
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3988  0.1985 -0.5186  0.5985  0.3360  0.1359 -0.2215  0.8147 -0.3913
#>   0.5039  0.6446 -0.4038 -0.5564  0.6591  0.3530  0.1268  0.2137  0.1091
#>   0.3832 -0.0034 -0.2492  0.2508  0.3232 -0.0309 -0.1968  0.6375 -0.0240
#> 
#> Columns 10 to 18  0.3943 -0.2702  0.1660 -0.4983  0.2311 -0.4457 -0.3052  0.3574  0.1565
#>  -0.2547  0.0508 -0.1432 -0.5691  0.6456 -0.0188  0.2690 -0.1811  0.1166
#>   0.0588 -0.7154  0.2524  0.1607  0.7278 -0.0452 -0.5068 -0.2114  0.2062
#> 
#> Columns 19 to 20 -0.5695 -0.2812
#>  -0.6577 -0.7361
#>  -0.2701 -0.3186
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1580  0.1281  0.3432 -0.2574  0.7993 -0.2550 -0.0830  0.3044  0.1013
#>   0.4412  0.3078 -0.1860 -0.1399  0.4296 -0.6181  0.3138  0.5400 -0.0725
#>   0.3158  0.2065 -0.0638 -0.2060  0.4476 -0.3411 -0.2177  0.3670 -0.1721
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8868  0.3644 -0.2533 -0.6597  0.5959  0.4784 -0.1705  0.5545  0.0685
#>  -0.4855  0.1375  0.5194 -0.3053 -0.5396 -0.6162 -0.2650 -0.1987 -0.6876
#>  -0.6281  0.3315 -0.4642 -0.2993  0.1616  0.1673  0.3594  0.5363 -0.2756
#> 
#> Columns 10 to 18 -0.8047 -0.3117  0.0432  0.2984  0.8866 -0.0132 -0.6523  0.7796  0.1656
#>  -0.5518  0.4556  0.5102  0.3191  0.4980  0.5046 -0.8971  0.8219  0.4031
#>  -0.4443  0.0075  0.3609 -0.2257  0.5282 -0.0461 -0.5213  0.1274  0.1771
#> 
#> Columns 19 to 20 -0.2210  0.1058
#>  -0.1377  0.0334
#>   0.1733  0.0968
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6328  0.1589 -0.3279  0.1503  0.6250 -0.2804 -0.1888  0.5422 -0.0187
#>   0.0715  0.1788 -0.4700 -0.1351  0.5494 -0.4711 -0.5523  0.3425 -0.0417
#>   0.1506 -0.3665 -0.2156 -0.0920  0.6516 -0.3059 -0.1769  0.6213  0.1312
#> 
#> Columns 10 to 18  0.5773 -0.6186 -0.4760 -0.0433 -0.2846 -0.0269 -0.1112 -0.2130  0.1473
#>   0.3126 -0.5970 -0.2847  0.3247 -0.3383 -0.2137 -0.0719 -0.1038  0.2710
#>   0.5069 -0.5272 -0.1253 -0.0526  0.0752 -0.0744 -0.1415 -0.0493 -0.0468
#> 
#> Columns 19 to 20 -0.1815 -0.4614
#>   0.2234 -0.0677
#>  -0.1379 -0.4458
#> [ CPUFloatType{2,3,20} ]
#>