RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7744 -0.1179  0.8441 -0.4269  0.4963  0.6301  0.3368  0.0153  0.8515
#>  -0.7148 -0.7362 -0.3829 -0.3642 -0.4000  0.0382  0.7503 -0.2551 -0.8129
#>   0.6680  0.0608 -0.4450  0.2135  0.6817  0.5612 -0.1240  0.7888  0.4689
#> 
#> Columns 10 to 18  0.3670  0.8280  0.4837  0.4383  0.4382 -0.5654  0.7135 -0.6272 -0.8063
#>   0.2254 -0.4801  0.1195 -0.6111 -0.4638 -0.3554  0.0120  0.4137 -0.0739
#>   0.5304  0.1644 -0.4053  0.4593  0.5701 -0.5181  0.7899  0.0225  0.3078
#> 
#> Columns 19 to 20  0.8152 -0.6826
#>   0.7629  0.7686
#>  -0.5391  0.6198
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1728 -0.6739  0.2881 -0.0363  0.4972 -0.3931  0.1518 -0.0420  0.5293
#>  -0.3333 -0.2894 -0.4776  0.1986 -0.1491 -0.2218  0.7757 -0.1679 -0.4901
#>  -0.5328  0.0149 -0.6353  0.0498  0.1126 -0.6709  0.2347  0.1101 -0.5952
#> 
#> Columns 10 to 18  0.5892  0.0419 -0.3638 -0.6218 -0.6691 -0.1076 -0.1052 -0.2410  0.2058
#>  -0.6150 -0.1219 -0.2628 -0.1538  0.1557  0.3770 -0.3859 -0.1088 -0.1474
#>   0.3731  0.5989 -0.2034  0.4538 -0.7483 -0.4607 -0.3933 -0.4895  0.0042
#> 
#> Columns 19 to 20  0.2808  0.4875
#>   0.5833  0.0778
#>   0.3597  0.2023
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4100  0.0568  0.1737 -0.1659  0.6132  0.2106 -0.2017 -0.6786  0.0045
#>   0.0011 -0.5908 -0.4288 -0.0001 -0.1524  0.1541  0.5695  0.2471  0.0744
#>   0.2175 -0.4821  0.0702 -0.0227  0.0676 -0.3273  0.0483  0.0219 -0.3783
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3921 -0.1487 -0.5257  0.4345 -0.3001  0.1388 -0.4217  0.0894 -0.3058
#>   0.4520 -0.2110 -0.2418 -0.5757  0.4932  0.2328 -0.1107 -0.4560 -0.1497
#>   0.0107 -0.1164 -0.7071 -0.3841  0.2780 -0.7003  0.0265 -0.4185 -0.5407
#> 
#> Columns 10 to 18  0.0460  0.3527  0.3380 -0.1395  0.5466  0.2934  0.2980  0.6045 -0.2527
#>   0.5348 -0.4111  0.4065  0.0090 -0.3829 -0.5728  0.7558  0.5642 -0.4099
#>   0.0286 -0.3822  0.5223  0.0761 -0.4282 -0.6479 -0.3630 -0.2705  0.1975
#> 
#> Columns 19 to 20 -0.2564 -0.1682
#>   0.0243 -0.3729
#>   0.5802 -0.7199
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1658  0.0258 -0.0240 -0.2132  0.1152  0.2963  0.3464 -0.3593 -0.0333
#>   0.2629 -0.5961  0.4132 -0.3225  0.2961 -0.0015  0.2561 -0.5156 -0.2975
#>   0.3458 -0.6073  0.5521 -0.1730  0.3183 -0.1227  0.4436 -0.1140 -0.1967
#> 
#> Columns 10 to 18  0.0177  0.3565 -0.5304 -0.1229 -0.2536 -0.4810 -0.2016 -0.3664 -0.5011
#>  -0.1457  0.2004  0.0589 -0.4020 -0.5922 -0.3346 -0.1440 -0.4299  0.1954
#>  -0.1208  0.1354  0.3050 -0.3178 -0.0459 -0.1491 -0.1109 -0.1192 -0.3399
#> 
#> Columns 19 to 20  0.5673  0.0908
#>   0.5777  0.2794
#>   0.2936 -0.2294
#> [ CPUFloatType{2,3,20} ]
#>