RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4133  0.1505 -0.5715  0.8163  0.1711 -0.0251  0.5797 -0.6675 -0.8188
#>   0.3149  0.3260 -0.6095  0.6453 -0.3291  0.7143  0.1289  0.0568  0.6075
#>  -0.4935  0.6342  0.5528 -0.1799  0.6310  0.3397  0.1397  0.0021  0.1790
#> 
#> Columns 10 to 18 -0.4427 -0.2103  0.0270  0.5905  0.3293  0.8625  0.4402 -0.3152 -0.0068
#>  -0.5471  0.6870 -0.2917 -0.4507  0.6784  0.2953 -0.5737 -0.6206 -0.7308
#>  -0.4369  0.1060 -0.0638 -0.0138  0.5694  0.5764 -0.6649 -0.3646 -0.6893
#> 
#> Columns 19 to 20 -0.9632  0.8325
#>  -0.1615  0.3793
#>  -0.6557  0.6673
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4687 -0.1639  0.0207  0.2834  0.1004 -0.1812 -0.2431 -0.5055  0.3255
#>  -0.0222  0.2232  0.1010  0.6430 -0.2792 -0.5800  0.0319  0.0380  0.5825
#>  -0.4609 -0.3491  0.0897  0.6711 -0.0171 -0.2108  0.0177 -0.3756  0.5652
#> 
#> Columns 10 to 18 -0.5300  0.1426  0.3093 -0.0535 -0.5847  0.3676  0.2792  0.0476 -0.5794
#>  -0.2955  0.2441 -0.2291 -0.2573  0.4611  0.4475 -0.0499 -0.1821 -0.1450
#>  -0.0591  0.1362 -0.1672 -0.4418  0.0256  0.2876  0.1644 -0.0915 -0.0399
#> 
#> Columns 19 to 20 -0.8789  0.5901
#>  -0.4990  0.6133
#>  -0.4961  0.3464
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0474 -0.1663  0.4109  0.4028 -0.1003 -0.5135 -0.1029 -0.3725  0.1634
#>   0.0837  0.2714  0.1951  0.3701 -0.1134 -0.0269  0.3911 -0.0558  0.7063
#>  -0.0300 -0.1244  0.2867  0.0602 -0.0239 -0.2575  0.2542 -0.1641  0.3874
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0807  0.8003 -0.3003 -0.0943  0.4919 -0.5439 -0.1380 -0.7427 -0.4234
#>   0.6687  0.1978 -0.1795  0.3449  0.0386 -0.3599  0.3720 -0.0691  0.1820
#>   0.4529  0.3660  0.6994  0.3133 -0.4273 -0.2709 -0.4389  0.5728  0.2742
#> 
#> Columns 10 to 18 -0.6246  0.1451 -0.2453  0.2424  0.6432  0.1622 -0.3074  0.0255  0.4541
#>   0.3381 -0.1768 -0.4498 -0.4022  0.3537 -0.3685 -0.3519  0.4018 -0.1964
#>  -0.1005 -0.5724  0.1633  0.6840  0.8808  0.2022 -0.9075  0.0414 -0.4244
#> 
#> Columns 19 to 20 -0.2712  0.6192
#>  -0.4626  0.1719
#>  -0.7688  0.4623
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1626 -0.0127  0.5907  0.7519 -0.2801 -0.5014  0.1716 -0.1495  0.6165
#>  -0.1678 -0.1509  0.0867  0.3143  0.1369 -0.3914  0.1558 -0.3195  0.2717
#>   0.3462  0.0119  0.1053  0.5731 -0.4192 -0.1715 -0.0483 -0.2572  0.4975
#> 
#> Columns 10 to 18 -0.2186  0.4178 -0.1357 -0.0268  0.0376 -0.3336 -0.0135 -0.5943 -0.5623
#>   0.1609 -0.2480 -0.2529 -0.1329 -0.3288  0.2385  0.6200 -0.2382 -0.2914
#>  -0.4419  0.3006  0.1918 -0.0340 -0.0835  0.1805  0.0679 -0.3953 -0.4868
#> 
#> Columns 19 to 20 -0.0202  0.5210
#>  -0.5266  0.5443
#>  -0.6051  0.2208
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>