RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0121 -0.0946 -0.0322 -0.4367  0.0999  0.2908 -0.8792  0.8208  0.0075
#>  -0.6034  0.7519  0.6429  0.3494  0.3281  0.9247 -0.5840  0.0086 -0.7157
#>   0.4655 -0.7573  0.6808 -0.4048 -0.2430  0.0136  0.5766  0.6038 -0.3028
#> 
#> Columns 10 to 18 -0.2963  0.7091 -0.8269 -0.6836 -0.5595 -0.4733  0.7327  0.4245  0.0980
#>   0.2012 -0.4777  0.1211 -0.6894 -0.5976  0.7474  0.1194  0.6760 -0.5683
#>  -0.7389  0.0456 -0.7387  0.3177 -0.7339  0.1937 -0.2102 -0.3185 -0.0308
#> 
#> Columns 19 to 20  0.3446  0.2144
#>  -0.3741 -0.6013
#>   0.9263  0.4822
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3242  0.5655  0.5529  0.2697 -0.0860  0.4218 -0.3064  0.0578 -0.1297
#>  -0.5137 -0.5847  0.2845 -0.4008 -0.5862 -0.3225  0.2151  0.2684 -0.7527
#>   0.1714  0.2683 -0.0333 -0.5505  0.2756  0.4816 -0.6202  0.0186 -0.4829
#> 
#> Columns 10 to 18 -0.5332 -0.4072 -0.3445  0.4717 -0.3049  0.1110 -0.5092  0.4472 -0.1619
#>  -0.1689 -0.3894  0.1262 -0.0844 -0.5324  0.5115 -0.5155  0.6982 -0.0897
#>  -0.0173 -0.3539  0.2272  0.5983 -0.0667 -0.2821 -0.5037  0.6150 -0.2355
#> 
#> Columns 19 to 20  0.1602 -0.4301
#>  -0.2768 -0.7021
#>  -0.1161 -0.7192
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3119  0.3495  0.3957  0.3449  0.3741  0.4831  0.0538 -0.5814 -0.1824
#>  -0.5278 -0.1136  0.2097 -0.1014 -0.2629  0.2694 -0.5500 -0.8342  0.1521
#>  -0.6936  0.4614  0.1945  0.2582 -0.2867  0.3026  0.0162 -0.5156 -0.1425
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6111  0.5259  0.5890 -0.4359  0.7706  0.1171  0.5541 -0.4308 -0.6698
#>  -0.5373 -0.4112  0.4340  0.2649  0.7423 -0.0398  0.0160 -0.6750 -0.8995
#>  -0.0147 -0.4762  0.4275 -0.1915  0.1944 -0.0099 -0.1414  0.2756 -0.1958
#> 
#> Columns 10 to 18  0.0289 -0.1088 -0.2521 -0.6111 -0.2698 -0.1724 -0.4166 -0.2722 -0.1452
#>   0.4268 -0.0660  0.1023 -0.5106 -0.4356 -0.4638 -0.6046 -0.4896  0.4920
#>   0.3485  0.3219  0.0450 -0.5717 -0.1243  0.5708 -0.1702 -0.7043  0.2145
#> 
#> Columns 19 to 20 -0.3613  0.7309
#>   0.0091 -0.2503
#>  -0.0749  0.3901
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3206  0.2898  0.4351  0.1024  0.1268  0.5135 -0.1064  0.0109 -0.5492
#>  -0.4941  0.3649  0.1518 -0.4002 -0.2032  0.2550 -0.5044  0.4503 -0.2223
#>  -0.5005  0.3014  0.4098 -0.1790  0.0356  0.0954  0.1169  0.2574 -0.1470
#> 
#> Columns 10 to 18 -0.2691  0.0477 -0.2840  0.0881 -0.5184  0.1863 -0.1389  0.5201  0.0274
#>  -0.3412 -0.3369 -0.5641  0.5492 -0.4277  0.1904 -0.4189 -0.1406  0.1022
#>  -0.0368 -0.2761 -0.4011  0.2962 -0.6096  0.0838 -0.0260  0.2894  0.0387
#> 
#> Columns 19 to 20  0.3976 -0.2357
#>   0.4475 -0.2266
#>   0.1144 -0.3823
#> [ CPUFloatType{2,3,20} ]
#>