RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8966  0.5246  0.5075 -0.3239  0.1856  0.1112  0.5113 -0.8988  0.1758
#>  -0.0737  0.7207 -0.4527  0.8832  0.5806  0.0593  0.4461  0.7638 -0.1801
#>   0.1685  0.2691  0.0958 -0.4605 -0.3498  0.1032  0.2413 -0.7955 -0.1593
#> 
#> Columns 10 to 18 -0.3549 -0.3306  0.2432 -0.2719 -0.4559 -0.8578 -0.6104  0.6757 -0.1182
#>   0.6652 -0.8774  0.0773  0.0554  0.4271  0.3580 -0.2550  0.2937 -0.1523
#>  -0.4892 -0.7876  0.4170 -0.3014  0.4583 -0.9014  0.3235  0.7585  0.0798
#> 
#> Columns 19 to 20  0.1604 -0.6926
#>  -0.7278 -0.7665
#>   0.8403 -0.7488
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2493 -0.1851 -0.6449  0.2857  0.1056  0.3405  0.0191 -0.1577 -0.3542
#>  -0.5093  0.7035 -0.3117 -0.0291 -0.2149  0.6639 -0.3086  0.3210 -0.5896
#>  -0.4681 -0.1884 -0.8859  0.4118  0.0748  0.0302  0.2717 -0.4635 -0.5454
#> 
#> Columns 10 to 18  0.0172  0.0752 -0.2307  0.3588  0.1764  0.3083  0.5552  0.4328 -0.1084
#>   0.6064 -0.2779 -0.5682  0.7214  0.7727  0.1229  0.1699  0.2345 -0.5082
#>   0.3143 -0.4491 -0.8157  0.1511  0.0381  0.0756 -0.2359  0.5109  0.3447
#> 
#> Columns 19 to 20 -0.2678 -0.5821
#>  -0.4284 -0.4278
#>   0.0709 -0.4411
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4303  0.6614 -0.0588 -0.1614  0.0640  0.4228  0.1505  0.3895 -0.3905
#>  -0.5262  0.6636  0.0463 -0.1146 -0.4283  0.5494  0.1083  0.5537 -0.2397
#>  -0.2305  0.4397 -0.3469  0.1401  0.0147  0.3009  0.4406  0.3396 -0.1920
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7384 -0.6898  0.5812  0.8517  0.3107 -0.2050  0.1750 -0.6821  0.2831
#>   0.3316 -0.5663 -0.6123  0.5180  0.5407 -0.0653 -0.5959 -0.5537  0.1213
#>   0.2393 -0.4010 -0.1574  0.3823  0.2256 -0.3848 -0.8409 -0.3921  0.3002
#> 
#> Columns 10 to 18 -0.5679 -0.4635 -0.6628 -0.4960  0.6240  0.6341 -0.5034 -0.1741  0.4141
#>   0.9267 -0.5787 -0.1016  0.3087 -0.2283 -0.4511 -0.6616  0.4271  0.1728
#>   0.7532 -0.1953 -0.5575 -0.1263 -0.2565 -0.1804  0.3438  0.6672 -0.5453
#> 
#> Columns 19 to 20 -0.4414  0.2122
#>  -0.2879  0.2260
#>  -0.2252  0.2767
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1891  0.4375 -0.5815  0.3767 -0.1698 -0.0101  0.2197  0.5966 -0.3488
#>   0.0242  0.3473 -0.4962  0.6118  0.3415  0.4321  0.1500  0.3370 -0.5801
#>  -0.5837  0.4370 -0.3175  0.2976  0.2748  0.5073 -0.0368  0.4354 -0.6911
#> 
#> Columns 10 to 18  0.4425  0.2763 -0.4182  0.2680  0.8200 -0.1704  0.1529  0.5563 -0.3516
#>   0.3733 -0.3328 -0.3832 -0.3190  0.5895 -0.4761  0.3234  0.4889 -0.2756
#>   0.4944 -0.0147 -0.7411  0.5060  0.3455 -0.1830  0.4844  0.5023 -0.1544
#> 
#> Columns 19 to 20 -0.1384 -0.5760
#>  -0.1698 -0.4649
#>   0.2018 -0.4704
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>