RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5454 -0.0480  0.8437  0.5164 -0.6959 -0.5535  0.4930 -0.2983  0.5400
#>  -0.2621 -0.1175  0.1821 -0.1674  0.2669 -0.2736  0.3763  0.6402  0.4564
#>  -0.1484 -0.9369 -0.1140  0.7759 -0.6997 -0.2672 -0.0056  0.7483 -0.3831
#> 
#> Columns 10 to 18  0.3203  0.6410  0.2783  0.3924 -0.2855  0.5136  0.6991  0.0132  0.4681
#>   0.1553 -0.7369  0.6054  0.2479 -0.4205 -0.2520  0.7580 -0.5757  0.6779
#>   0.2467  0.9194 -0.9302 -0.9319 -0.5925  0.5360 -0.6657 -0.1935 -0.2195
#> 
#> Columns 19 to 20 -0.0824 -0.2003
#>  -0.5162 -0.1332
#>  -0.1122 -0.0549
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1552  0.0796 -0.6252  0.1699  0.4375 -0.1131  0.3213  0.4480 -0.5602
#>  -0.5885 -0.6706 -0.5886  0.1479  0.0229  0.3117  0.2096 -0.0949 -0.3199
#>   0.2428 -0.4432 -0.0626  0.0183 -0.5160 -0.5739 -0.5136  0.4024 -0.1452
#> 
#> Columns 10 to 18 -0.3413 -0.2091  0.2290 -0.0628 -0.6010  0.0319 -0.0331 -0.1738 -0.4162
#>   0.1751 -0.0438 -0.0802 -0.6649 -0.8306  0.2194  0.4120  0.2012 -0.3502
#>  -0.4845  0.3283 -0.1222 -0.0630 -0.5263  0.0409 -0.0544  0.0582 -0.5716
#> 
#> Columns 19 to 20 -0.4317  0.1259
#>  -0.0484  0.0390
#>  -0.2492  0.1891
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0468 -0.6345 -0.3631  0.1380 -0.5066 -0.2671 -0.2667  0.5775 -0.3027
#>   0.1932 -0.3271 -0.1583  0.1196  0.0538  0.4064 -0.2479  0.1702 -0.5438
#>  -0.0610 -0.3629 -0.3634  0.1615 -0.6581 -0.5715  0.1054  0.0141  0.0108
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1926 -0.3452 -0.4782  0.7698  0.4041  0.2751 -0.2208  0.0935  0.0734
#>   0.4093 -0.7589  0.0003  0.6661  0.3730  0.0804  0.0093  0.1197  0.1298
#>  -0.5912 -0.0503  0.0170 -0.1378 -0.1386  0.4254  0.0643 -0.3921 -0.5265
#> 
#> Columns 10 to 18 -0.3091  0.4797 -0.3288  0.4575  0.1177  0.0147 -0.2060  0.3175  0.2979
#>  -0.8004  0.4468  0.7527 -0.0408 -0.1735  0.6334  0.3861  0.6479  0.7002
#>  -0.7897  0.0109 -0.1941  0.8166 -0.1676  0.0149  0.3686 -0.1754  0.1738
#> 
#> Columns 19 to 20  0.3699  0.0919
#>  -0.4376 -0.4139
#>   0.0814 -0.6844
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1224 -0.4933 -0.2953 -0.0464 -0.6876 -0.1599 -0.4100  0.1805 -0.2387
#>   0.2431 -0.2645 -0.5768 -0.2272 -0.7012 -0.5947 -0.4359  0.5730  0.1840
#>   0.2838 -0.3342 -0.2496 -0.1087 -0.5301 -0.0853 -0.3377  0.0777 -0.3195
#> 
#> Columns 10 to 18 -0.1346  0.2470  0.0567  0.0471 -0.5134  0.1613  0.4364  0.1214 -0.2358
#>  -0.1314 -0.2493  0.0708  0.2883 -0.7170  0.0047  0.5635  0.0047  0.2558
#>   0.0127 -0.0326  0.3035 -0.1099 -0.5074  0.1534  0.3060  0.2176 -0.1791
#> 
#> Columns 19 to 20  0.0834 -0.2758
#>  -0.5778 -0.1033
#>   0.1506 -0.3728
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>