RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0278  0.1483  0.6189 -0.4008 -0.4143  0.4779  0.2955  0.0321 -0.8148
#>   0.5663 -0.0659 -0.0209 -0.1985 -0.1348  0.7789  0.4806 -0.1873 -0.0077
#>   0.7898 -0.1627  0.7449 -0.0094  0.7585  0.8016 -0.7442 -0.3799 -0.8271
#> 
#> Columns 10 to 18  0.4814 -0.0145 -0.6465  0.3744 -0.4805 -0.8814 -0.6913 -0.5961 -0.6435
#>   0.3984  0.2814 -0.2456 -0.0334  0.7505 -0.4398 -0.3553  0.0772  0.0184
#>   0.2509  0.1713  0.0751  0.3106 -0.6562 -0.7480 -0.9482  0.5956 -0.3798
#> 
#> Columns 19 to 20 -0.5717 -0.0538
#>   0.9600 -0.0085
#>  -0.8305  0.2355
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4170  0.3599  0.1291 -0.5717  0.4833  0.3091  0.3308 -0.1509 -0.5598
#>   0.0977  0.1053  0.2502 -0.3048  0.5972 -0.0070  0.6650 -0.0455 -0.4832
#>  -0.5611 -0.1084  0.1219 -0.3885 -0.0735  0.0999  0.4053  0.1708  0.0589
#> 
#> Columns 10 to 18  0.2510  0.2334 -0.0808  0.4315 -0.7269 -0.2554 -0.6599  0.1953 -0.2441
#>   0.0796  0.4621 -0.0176  0.1944  0.0294 -0.0804 -0.1292  0.1643  0.0048
#>   0.4761 -0.0825 -0.4135  0.5101 -0.6668 -0.2104 -0.0967 -0.1027 -0.7053
#> 
#> Columns 19 to 20 -0.1612  0.1754
#>  -0.0432 -0.0483
#>   0.2272  0.5528
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2062 -0.1856 -0.0065 -0.1040  0.1934 -0.3543  0.3127  0.2889 -0.4631
#>  -0.0706 -0.5950 -0.2598 -0.3769  0.0068 -0.0555  0.4186  0.0203  0.2285
#>   0.5761  0.1392  0.2535 -0.4977  0.4663  0.4477  0.2786 -0.2081 -0.1210
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3385  0.2866  0.8925 -0.2058  0.0010 -0.7672  0.2706  0.4695  0.3198
#>   0.2467  0.4387  0.0274 -0.8258 -0.3833 -0.6904 -0.1732  0.5974  0.6466
#>  -0.3749  0.1616 -0.0802 -0.0848 -0.6644 -0.4440 -0.5104  0.5094  0.0899
#> 
#> Columns 10 to 18 -0.5779  0.2298 -0.1642  0.4421 -0.4037  0.5337 -0.3314  0.6249 -0.1795
#>  -0.6688  0.3028 -0.3593  0.4468  0.4316  0.0488 -0.1569  0.5638 -0.3985
#>  -0.2540 -0.0377 -0.3255 -0.3465  0.4444  0.5420  0.5867  0.8000 -0.7436
#> 
#> Columns 19 to 20  0.2712  0.1684
#>  -0.6002  0.2108
#>  -0.3843 -0.4283
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4396  0.3867  0.4149 -0.0184  0.3153 -0.0415  0.4769  0.1430 -0.3581
#>   0.3943 -0.0240 -0.1240 -0.5434  0.0208 -0.0447  0.5355  0.1756 -0.5562
#>   0.1727 -0.0706 -0.1370 -0.3890  0.5334 -0.2975  0.5281  0.0526 -0.3319
#> 
#> Columns 10 to 18  0.0519  0.1458 -0.1066  0.2506  0.0921 -0.0635 -0.4217 -0.2086  0.0536
#>   0.0903  0.1728 -0.2738 -0.2861 -0.3272 -0.0107 -0.4425 -0.2616 -0.3874
#>   0.3425  0.2787 -0.2614  0.6307 -0.3627 -0.5280 -0.1588 -0.1841 -0.6212
#> 
#> Columns 19 to 20 -0.0157  0.4115
#>   0.1841  0.6221
#>  -0.1277  0.0043
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>