RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6561 -0.7909  0.0886  0.1661 -0.4820 -0.4902  0.6829 -0.4348  0.5524
#>   0.5603 -0.1576  0.7117  0.5251  0.1643  0.6764 -0.4272  0.6258 -0.1674
#>  -0.2308 -0.8450 -0.2769 -0.1290 -0.3998  0.2827  0.2775  0.7473  0.3571
#> 
#> Columns 10 to 18 -0.4195 -0.1066 -0.3322 -0.7427 -0.6390  0.3358 -0.5693 -0.0254 -0.0030
#>  -0.2263  0.5080 -0.1762  0.3663 -0.3495 -0.0850  0.3792 -0.5532  0.6386
#>   0.2655  0.2395 -0.7987  0.6231 -0.5163 -0.9350  0.0160 -0.5075  0.8871
#> 
#> Columns 19 to 20  0.0205 -0.1933
#>   0.4016  0.1269
#>   0.5444  0.7940
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1612 -0.3211 -0.0095 -0.3890 -0.1951  0.3138 -0.6777  0.1480 -0.4809
#>   0.2343  0.0044  0.0548  0.1635 -0.0692 -0.2279  0.1896 -0.4578 -0.1584
#>   0.1010 -0.0357 -0.1070  0.2907  0.2870 -0.0440 -0.2570 -0.1461 -0.3585
#> 
#> Columns 10 to 18  0.1381 -0.4830 -0.2372  0.1648 -0.6020  0.0209 -0.4117  0.0712 -0.3560
#>  -0.2109 -0.3841 -0.1258  0.3763  0.0790  0.3750 -0.0534  0.1747 -0.6905
#>   0.4241 -0.5309 -0.2824 -0.2343 -0.7051  0.7556  0.0506  0.1186 -0.4956
#> 
#> Columns 19 to 20 -0.6553  0.1238
#>  -0.0023  0.1528
#>  -0.7103 -0.4840
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5655 -0.3427 -0.3084  0.0922  0.0447  0.0312 -0.2010  0.4309  0.1129
#>  -0.1308  0.3570 -0.3590 -0.2867  0.2670  0.0243 -0.0426  0.0463  0.4062
#>  -0.3178 -0.3484  0.0779  0.0792 -0.5419  0.2678 -0.4837  0.5253  0.7653
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2432 -0.0926  0.7103 -0.2906 -0.9372 -0.8701 -0.1380 -0.2564 -0.1811
#>  -0.0180 -0.0490  0.0431  0.3162 -0.3626 -0.1912  0.2385  0.3884  0.4008
#>  -0.7797  0.7059 -0.3566 -0.4088 -0.6744 -0.0738  0.2002  0.1996 -0.2283
#> 
#> Columns 10 to 18 -0.3926 -0.3248  0.6700  0.8390  0.8316 -0.9466  0.7673  0.1873 -0.4828
#>  -0.6434 -0.3042 -0.5571  0.5369  0.6177 -0.5570  0.1290  0.1815  0.4050
#>  -0.5437  0.1990  0.4243  0.4258  0.3924 -0.5539  0.6243 -0.3557  0.2283
#> 
#> Columns 19 to 20 -0.2097  0.6690
#>  -0.3300 -0.4521
#>  -0.0333  0.3391
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1101 -0.0689 -0.0466  0.4144  0.3688 -0.5486  0.2450  0.1576  0.1625
#>   0.3520 -0.2526  0.1606 -0.0885 -0.0343  0.1816 -0.1463  0.0652 -0.1462
#>  -0.0409  0.0509 -0.1658  0.1324  0.1153 -0.1219 -0.0843  0.4948  0.6512
#> 
#> Columns 10 to 18 -0.0472 -0.0612 -0.4825  0.4147 -0.1530 -0.0497 -0.0447  0.1457 -0.4828
#>   0.1039 -0.1844 -0.2158  0.5756 -0.1689  0.3594 -0.3597  0.2633 -0.1617
#>   0.1101  0.3420  0.1596  0.2649 -0.1471  0.3286  0.1600 -0.2305  0.2848
#> 
#> Columns 19 to 20 -0.2300 -0.3572
#>  -0.2324 -0.0721
#>  -0.0642 -0.4928
#> [ CPUFloatType{2,3,20} ]
#>