RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0251 -0.6299 -0.1933 -0.2359  0.6456  0.0045 -0.2696  0.0011 -0.1172
#>   0.1759  0.5919 -0.4249  0.7805  0.6728  0.4825 -0.8290 -0.0753 -0.2498
#>   0.2944 -0.6334  0.6991 -0.3811 -0.2385 -0.6200 -0.7516  0.0195 -0.0302
#> 
#> Columns 10 to 18 -0.0278 -0.7611  0.7191  0.4981 -0.2780 -0.1759 -0.6399  0.0499 -0.0639
#>   0.9490 -0.3488  0.6823  0.0368  0.4041  0.3055  0.4243 -0.1561  0.4365
#>  -0.0159  0.3210  0.6646  0.7083  0.3500  0.6047 -0.0300 -0.4015 -0.5057
#> 
#> Columns 19 to 20  0.1339 -0.6086
#>   0.6901 -0.0342
#>  -0.5214 -0.3976
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0905 -0.1762 -0.1543  0.1298 -0.5517 -0.3325 -0.2045  0.1448  0.1506
#>   0.3351  0.0189  0.2612 -0.2669 -0.2564 -0.0117 -0.6159 -0.3392 -0.1133
#>  -0.2884 -0.5820  0.4722 -0.0272 -0.2374 -0.5963 -0.3276  0.2142  0.2187
#> 
#> Columns 10 to 18  0.0558  0.3736  0.4355  0.5275  0.3102 -0.0744 -0.0468  0.0118 -0.3128
#>  -0.0110  0.3816  0.1658  0.4501  0.6834  0.3542  0.1353  0.3343 -0.5727
#>  -0.2640  0.0807  0.1555  0.1587  0.3119 -0.3364 -0.4712 -0.5303 -0.5542
#> 
#> Columns 19 to 20  0.0155  0.2881
#>  -0.4759 -0.1541
#>  -0.5207  0.3927
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2751 -0.0465  0.0789 -0.0156 -0.5516 -0.4577 -0.0878 -0.1478 -0.0712
#>  -0.3385 -0.0450  0.0051  0.2419 -0.2133 -0.5666 -0.0474  0.0139 -0.1428
#>  -0.4725  0.1076  0.2126 -0.2778 -0.6469 -0.4309 -0.6184  0.5804  0.2829
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1315 -0.6268  0.5563 -0.1185 -0.1659 -0.3497 -0.4979  0.3354  0.3025
#>  -0.2339 -0.5177  0.2182  0.0757  0.2556 -0.6370 -0.8559 -0.3033  0.0461
#>   0.2627 -0.0691  0.2274  0.0460 -0.0869 -0.0347  0.0849  0.1265 -0.4517
#> 
#> Columns 10 to 18  0.2786 -0.2384  0.0468 -0.3223  0.5605 -0.3493  0.4933 -0.0084 -0.1421
#>  -0.5506 -0.0131  0.1273 -0.6557  0.2670 -0.1489 -0.0714  0.2647 -0.1257
#>   0.7635 -0.7839  0.4722  0.9119  0.9019 -0.1874 -0.1510  0.4356  0.7020
#> 
#> Columns 19 to 20  0.3633  0.6972
#>   0.0615  0.3565
#>  -0.8197  0.7417
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3036 -0.3960  0.4347  0.0863 -0.2612 -0.4368 -0.2723 -0.4349 -0.0319
#>   0.0487 -0.1059  0.2645  0.3025 -0.2882 -0.4222 -0.2668 -0.0544  0.0045
#>   0.1537 -0.6912  0.4156 -0.2664 -0.6896 -0.5227 -0.3094 -0.1056 -0.3661
#> 
#> Columns 10 to 18 -0.0073  0.5177  0.4425  0.0909  0.3316 -0.0113  0.1187 -0.2791 -0.3664
#>   0.0702  0.3907  0.2843  0.1138  0.2710 -0.1968 -0.0168 -0.4223 -0.2821
#>   0.2998  0.1846  0.7396  0.6944  0.2030  0.0373 -0.2686 -0.0522 -0.5707
#> 
#> Columns 19 to 20 -0.2996 -0.5347
#>  -0.1515 -0.2320
#>   0.2345  0.4124
#> [ CPUFloatType{2,3,20} ]
#>