RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1118  0.1388 -0.4493 -0.5998  0.1644 -0.5471  0.3763  0.0645  0.0096
#>   0.3734  0.9268  0.1249 -0.6840  0.3020 -0.7385  0.6147  0.3332 -0.9473
#>   0.7305  0.9305  0.4759 -0.7089  0.0679  0.1083  0.7700  0.3548  0.7766
#> 
#> Columns 10 to 18 -0.6209 -0.1293  0.6826  0.5198 -0.8281  0.6813  0.2726  0.9438  0.9581
#>   0.4745  0.3824  0.4195  0.1242 -0.1342  0.8568 -0.0932 -0.4324 -0.6939
#>   0.3548  0.1264 -0.1541  0.5319 -0.5701  0.8331 -0.4041 -0.5925 -0.6936
#> 
#> Columns 19 to 20 -0.1984  0.3426
#>   0.7057  0.8958
#>   0.6833  0.7526
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2405  0.7217  0.0686 -0.3818  0.3994 -0.2027 -0.5237  0.7418  0.2733
#>  -0.1754  0.7838 -0.0090 -0.2139  0.0672 -0.3164  0.3602 -0.0418  0.0163
#>  -0.2439  0.7646  0.3454 -0.3747  0.1035  0.0230  0.0423  0.0169 -0.5903
#> 
#> Columns 10 to 18 -0.2678  0.1542  0.5395 -0.2066 -0.2326  0.3858 -0.3395  0.7041  0.4875
#>   0.2745  0.1444  0.4547  0.4609 -0.3359  0.7764  0.4491 -0.1458  0.0917
#>   0.7098  0.0230  0.2361 -0.4281 -0.3707  0.7691  0.4787  0.2653 -0.1320
#> 
#> Columns 19 to 20  0.3966  0.3704
#>   0.1359  0.1867
#>   0.6432  0.2241
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3774  0.6030  0.0840 -0.4924  0.2718 -0.1570 -0.0430  0.5545  0.6793
#>  -0.3326  0.6386  0.3183 -0.3414  0.2388 -0.1172  0.0861  0.5602  0.0710
#>   0.2656  0.7721  0.0298 -0.0779  0.4024 -0.4067  0.2281  0.6567  0.1366
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5029  0.2899 -0.4544  0.0902 -0.2792 -0.3998  0.1288 -0.2516  0.8580
#>   0.0320  0.6345 -0.2339  0.1228 -0.0615  0.0407  0.4165 -0.2850  0.7386
#>   0.3229  0.3198 -0.1500 -0.2045 -0.4368 -0.3714  0.3145 -0.2809 -0.1111
#> 
#> Columns 10 to 18  0.4024 -0.4644 -0.1720  0.7588 -0.2207  0.3141  0.1223  0.6172 -0.0630
#>   0.0936 -0.5626 -0.5469  0.4478 -0.3739 -0.0010  0.1087 -0.2355 -0.1732
#>  -0.1051 -0.7059 -0.4492  0.3523 -0.2319  0.2108  0.0592 -0.3944 -0.3823
#> 
#> Columns 19 to 20 -0.1241 -0.7147
#>  -0.0864 -0.2473
#>  -0.1973  0.2183
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5678  0.4380  0.0396 -0.3644 -0.0769 -0.0880 -0.3309  0.4663  0.5566
#>  -0.3389  0.3140 -0.1240 -0.6624 -0.3383 -0.3353 -0.0722  0.2688  0.3984
#>  -0.0386  0.5169  0.1977 -0.1354 -0.3381 -0.1205  0.1914 -0.1827  0.1346
#> 
#> Columns 10 to 18 -0.0063  0.1724 -0.4000  0.2374 -0.0805  0.2453 -0.0662  0.6298  0.4645
#>  -0.2566  0.0018 -0.2841  0.3092 -0.2386  0.2613 -0.1353  0.4982  0.5591
#>   0.4274 -0.1050  0.1063  0.1890 -0.0830  0.4312  0.1655  0.3385  0.6160
#> 
#> Columns 19 to 20  0.4201  0.3515
#>   0.3465  0.2785
#>   0.4318  0.3499
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>