RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3688 -0.4668 -0.3090 -0.6530  0.5173 -0.2980  0.3569  0.7561 -0.4078
#>   0.1088 -0.3299  0.0005  0.3962  0.8410  0.3846 -0.5296 -0.4266 -0.5375
#>   0.0915  0.9118 -0.3427  0.3985  0.2630 -0.1645  0.6218 -0.3762 -0.0237
#> 
#> Columns 10 to 18 -0.2545  0.3864  0.1615 -0.5894 -0.4460 -0.3992  0.8330 -0.0166 -0.4392
#>   0.2557  0.1633  0.3844  0.5334 -0.1314 -0.3419  0.6134  0.1182  0.0242
#>  -0.7552 -0.8226  0.6698  0.1403  0.1137  0.7825  0.6356  0.4773  0.4851
#> 
#> Columns 19 to 20 -0.6506 -0.0171
#>   0.0801 -0.3200
#>  -0.7958 -0.6160
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0655 -0.2592 -0.0542 -0.0965  0.6402 -0.0664 -0.6709  0.6248 -0.0963
#>   0.0892  0.1179  0.0503  0.0718  0.5593 -0.0405 -0.4760  0.3768  0.4734
#>  -0.0478  0.1885 -0.6251 -0.2005  0.7642 -0.0784 -0.1153  0.1202 -0.3156
#> 
#> Columns 10 to 18 -0.3477 -0.2994 -0.2192 -0.0534 -0.0872 -0.6514  0.5582  0.1290 -0.1985
#>  -0.2053 -0.1858 -0.0114 -0.1832  0.1734 -0.0188  0.5189  0.0918 -0.5010
#>  -0.3924  0.1401  0.2095 -0.7029 -0.1802 -0.0032  0.8327  0.6698  0.1384
#> 
#> Columns 19 to 20  0.5391  0.2691
#>  -0.3804 -0.3731
#>  -0.1971 -0.4862
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2730 -0.4283  0.1854  0.2208  0.6391 -0.1380 -0.3978  0.5036  0.2490
#>  -0.3010 -0.1325 -0.3409 -0.1911  0.5804 -0.2358 -0.2726  0.3433 -0.0001
#>   0.3452  0.1080 -0.4586 -0.3291  0.9285 -0.3970 -0.5809  0.0812 -0.0349
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0251 -0.4661 -0.6343  0.4078 -0.3917  0.6350  0.5765 -0.0057 -0.1868
#>  -0.4930  0.3459 -0.1618 -0.0434  0.3633  0.3188  0.0339 -0.2157  0.2084
#>  -0.6338 -0.3744  0.0316  0.7004 -0.0411  0.3464 -0.5007 -0.6257  0.0295
#> 
#> Columns 10 to 18  0.2770 -0.1026 -0.0082  0.5715  0.3310  0.1318  0.3829 -0.1859 -0.5660
#>   0.5987 -0.1993 -0.0830 -0.0666 -0.1573 -0.1510 -0.1741  0.4102  0.2426
#>   0.7192  0.0710 -0.5837  0.1998 -0.1823 -0.4862  0.2885  0.1310 -0.2266
#> 
#> Columns 19 to 20 -0.0441 -0.7058
#>  -0.4943 -0.3825
#>  -0.1043  0.1864
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2428 -0.1097 -0.1830  0.1678  0.4528 -0.1539 -0.2518 -0.3231  0.2699
#>  -0.2348 -0.2667  0.0107  0.0529  0.4756 -0.2245 -0.3987  0.6121  0.1879
#>  -0.1540  0.1251  0.1610 -0.1055  0.0223 -0.1425 -0.6567  0.3469  0.4357
#> 
#> Columns 10 to 18 -0.0421 -0.3303 -0.1502 -0.4544  0.0430 -0.1059  0.1941 -0.0666  0.1778
#>  -0.4475 -0.4629 -0.2530 -0.2016  0.2054 -0.4343  0.5975 -0.1021 -0.2064
#>  -0.2715 -0.6776  0.1134 -0.4311  0.0868 -0.2685  0.3537 -0.2890 -0.0135
#> 
#> Columns 19 to 20  0.2070 -0.2415
#>   0.2069 -0.2949
#>   0.2512 -0.1495
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>