RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9429 -0.7155 -0.2826 -0.5893  0.2503 -0.6856  0.5250 -0.8096 -0.6758
#>  -0.1276 -0.5632 -0.0170 -0.5437 -0.0277 -0.4384  0.3899  0.2952 -0.1947
#>   0.0212  0.0977 -0.1685 -0.4547  0.0674 -0.2391 -0.0647 -0.7601 -0.5934
#> 
#> Columns 10 to 18  0.0683  0.5112  0.7930  0.7694 -0.8304  0.3198 -0.6662  0.6040  0.3142
#>   0.0500 -0.3380  0.2149 -0.0041 -0.1624 -0.3456 -0.6392  0.4680  0.0378
#>   0.5961 -0.6239 -0.2152 -0.6884  0.4545 -0.1409  0.1174  0.0699 -0.4773
#> 
#> Columns 19 to 20 -0.7180 -0.5743
#>  -0.6398 -0.0255
#>  -0.2014 -0.3050
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0838  0.0158  0.5959 -0.2929  0.2600  0.0248  0.4732  0.1229  0.0227
#>   0.2677  0.1458  0.1926 -0.3520 -0.0843  0.3500  0.1128  0.1915 -0.5303
#>  -0.1876 -0.7279  0.0570 -0.5687  0.5692 -0.2837  0.5391  0.0593  0.1104
#> 
#> Columns 10 to 18 -0.7642 -0.5764  0.4118 -0.2834 -0.4577 -0.0996 -0.4150  0.1767 -0.4881
#>   0.0531  0.2617  0.3203  0.3515  0.2313 -0.0505 -0.2127  0.6179  0.2168
#>   0.2259  0.0922 -0.0003  0.1046 -0.0429 -0.0085  0.2939  0.0951  0.2319
#> 
#> Columns 19 to 20  0.2968  0.0064
#>   0.4186 -0.3316
#>   0.4119  0.6075
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2326  0.0388 -0.0066 -0.2610  0.1065  0.1053  0.0094  0.2826  0.2348
#>  -0.5241 -0.2771  0.1785 -0.6846  0.4358 -0.5428  0.1053  0.4244  0.1106
#>   0.2652  0.3582  0.2018  0.1987 -0.0640  0.4716  0.0714 -0.2786  0.0328
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2569 -0.0110 -0.2510 -0.3062 -0.2457  0.5562  0.4704  0.0663 -0.7377
#>   0.1745 -0.8431 -0.7575  0.1291  0.5204  0.4135  0.1441 -0.3167 -0.2454
#>  -0.0922  0.0779 -0.0855 -0.2661 -0.1559  0.8463  0.1942  0.0276 -0.6759
#> 
#> Columns 10 to 18 -0.1794 -0.2745 -0.0330 -0.5328 -0.1561  0.3866 -0.2385 -0.1946 -0.2397
#>   0.0942 -0.3993  0.4359 -0.0745 -0.4776 -0.0673  0.0314 -0.3869 -0.9170
#>   0.1416  0.2079 -0.0701  0.1246 -0.4002  0.5262 -0.5297  0.0060 -0.0229
#> 
#> Columns 19 to 20  0.6473  0.3602
#>   0.0585  0.8774
#>   0.5350  0.0482
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3478 -0.0909  0.4525 -0.1804  0.5668  0.1161  0.1389  0.0963 -0.2602
#>  -0.1385  0.0960 -0.0219 -0.3081 -0.1333 -0.3182  0.0495  0.5878 -0.3043
#>  -0.2430 -0.0396  0.4902  0.0979  0.7079  0.5262  0.1654  0.1579  0.0148
#> 
#> Columns 10 to 18 -0.2893 -0.1739  0.3858  0.2582 -0.4005 -0.4632 -0.1881 -0.1507  0.1366
#>  -0.1841 -0.3060  0.4242  0.2496 -0.2731  0.1187 -0.6969 -0.2800 -0.2952
#>   0.2564  0.0800  0.4167  0.2359 -0.2549 -0.1786  0.3454 -0.2484  0.4428
#> 
#> Columns 19 to 20  0.0025  0.3777
#>   0.1660  0.1710
#>   0.0919  0.1662
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>