RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4570  0.4334  0.6497  0.3831  0.2819 -0.5450 -0.0932 -0.0636 -0.8201
#>   0.7068 -0.7619  0.0683 -0.6156 -0.9251  0.1468 -0.4347  0.1178 -0.2265
#>   0.4764  0.1732  0.0540 -0.4746 -0.7347  0.6252 -0.4234 -0.4820 -0.8152
#> 
#> Columns 10 to 18  0.4579 -0.1339  0.7333 -0.5355  0.2854  0.8034 -0.0601 -0.6086  0.9657
#>  -0.7189  0.4260  0.3723 -0.7838  0.8321  0.5216 -0.4143  0.5436 -0.9373
#>  -0.5721  0.7329  0.3608 -0.9268  0.5708  0.5141 -0.1347  0.0640  0.3060
#> 
#> Columns 19 to 20  0.3086  0.9290
#>   0.8405 -0.9442
#>   0.6290 -0.3642
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1886 -0.2260 -0.2199  0.0155 -0.7544  0.1953 -0.0462 -0.1928  0.2264
#>   0.6463 -0.2169  0.7589  0.5935 -0.5191  0.0875  0.3122  0.1887 -0.3695
#>   0.4376  0.3213  0.6578  0.2892 -0.3961  0.1838  0.2351  0.0932 -0.2119
#> 
#> Columns 10 to 18 -0.2221  0.6512 -0.1206 -0.4613  0.2052  0.5221  0.2627 -0.2259  0.4606
#>  -0.2759  0.0814 -0.2799 -0.0883  0.7942  0.2726 -0.0725  0.2806 -0.2678
#>   0.5395 -0.1170 -0.0944  0.0753  0.2429  0.0735  0.2879 -0.1410  0.1423
#> 
#> Columns 19 to 20  0.5859  0.5450
#>  -0.2861 -0.0022
#>  -0.5004 -0.2128
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0547 -0.4968 -0.4025  0.5186 -0.5359 -0.3472  0.2499 -0.2914 -0.1307
#>   0.4338  0.2585  0.1982  0.2748 -0.1987  0.0965 -0.1880  0.4915 -0.0090
#>   0.2762  0.3141  0.0248 -0.3608  0.0450  0.1288 -0.0106  0.0644 -0.0805
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0647 -0.1251 -0.3794  0.1436 -0.5329  0.3143 -0.1385  0.3200 -0.4397
#>  -0.5163 -0.3770  0.2894  0.1092 -0.1156 -0.5133  0.0381 -0.0724 -0.7173
#>  -0.4718  0.5222 -0.3711 -0.2247  0.3771 -0.3741  0.2221 -0.6828 -0.4707
#> 
#> Columns 10 to 18 -0.7059  0.1967 -0.1379 -0.0989  0.1077  0.1637  0.3319 -0.0145  0.1835
#>  -0.7967  0.7435 -0.6020  0.5807 -0.5225  0.5121 -0.6192  0.0528 -0.4836
#>  -0.6056  0.8850 -0.4904  0.1492 -0.4605  0.6294  0.6016 -0.7436 -0.6362
#> 
#> Columns 19 to 20  0.0197  0.2016
#>   0.4905 -0.8111
#>   0.5853 -0.4277
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0666 -0.4374 -0.5854  0.1121 -0.3783  0.1034  0.1841 -0.2001 -0.0478
#>   0.0106 -0.0157  0.0090  0.3614 -0.0860 -0.3252 -0.1614  0.2240 -0.0946
#>   0.0284  0.1236 -0.3266 -0.2046 -0.2423 -0.3205 -0.4154 -0.1517  0.1376
#> 
#> Columns 10 to 18 -0.2015  0.4129 -0.2474 -0.1802  0.2298  0.3110  0.4915  0.0746 -0.0868
#>   0.1286  0.4476 -0.2126  0.3492 -0.5232  0.1824  0.3921  0.2634  0.3154
#>   0.2387 -0.1982 -0.2765 -0.0621 -0.3007 -0.3292  0.3632 -0.4795  0.2965
#> 
#> Columns 19 to 20 -0.0700  0.4559
#>  -0.4756  0.2654
#>  -0.4653 -0.3269
#> [ CPUFloatType{2,3,20} ]
#>