RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6607  0.0863 -0.1687  0.8786 -0.6689  0.1260 -0.7310 -0.7013 -0.7654
#>   0.8149 -0.2748 -0.6762  0.4512  0.5437 -0.0387 -0.5006 -0.0954 -0.1238
#>   0.4432  0.1860 -0.6931  0.0206  0.5566  0.5791  0.5576 -0.6253  0.3309
#> 
#> Columns 10 to 18  0.6210 -0.8894  0.8433  0.5226 -0.4655 -0.6327 -0.2903  0.3801  0.7830
#>   0.4337  0.5292  0.8938 -0.4197 -0.7673 -0.8578 -0.4698 -0.4143  0.3374
#>  -0.1983  0.3745 -0.6409  0.1539 -0.7582 -0.8323 -0.7730  0.8432  0.6042
#> 
#> Columns 19 to 20 -0.3310  0.5118
#>  -0.1547 -0.5558
#>  -0.7569  0.2931
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3103 -0.3516 -0.4812 -0.3180  0.5895  0.1012  0.2273  0.2966 -0.5838
#>  -0.1308 -0.2552 -0.3529 -0.4258 -0.2163 -0.2079  0.1934 -0.6201  0.0779
#>   0.5817  0.0884  0.3970  0.3935  0.4446 -0.0247 -0.2500 -0.2986  0.2270
#> 
#> Columns 10 to 18 -0.1936  0.7047  0.1937 -0.6331 -0.6490 -0.4810 -0.6565  0.0618  0.4316
#>  -0.4553 -0.0895  0.0887  0.1972 -0.3326 -0.4622  0.3821  0.5021  0.5680
#>   0.1750 -0.0872  0.3345  0.0909 -0.3636 -0.1694  0.0368  0.3653  0.3386
#> 
#> Columns 19 to 20  0.1857  0.6517
#>   0.7470  0.3358
#>   0.2441 -0.1102
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1157 -0.5597  0.1031 -0.1326  0.1234  0.2980  0.4517 -0.1032 -0.1449
#>   0.3946  0.1158 -0.0659  0.3878  0.0283  0.4675 -0.0806 -0.0367 -0.4669
#>   0.4221 -0.4304 -0.3618  0.3132  0.2620  0.2556 -0.0998 -0.0125 -0.3653
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0382 -0.1532  0.5934  0.5377  0.2570  0.3835  0.1913 -0.4276  0.1459
#>  -0.4090 -0.0784 -0.3695  0.2235 -0.3681  0.3339  0.1785 -0.4748  0.3152
#>  -0.2469 -0.4447  0.7798  0.3125  0.4650  0.0242  0.2244  0.0065 -0.2810
#> 
#> Columns 10 to 18  0.0083  0.6639 -0.3804 -0.7560  0.6578  0.2443 -0.5296  0.5617  0.3556
#>   0.2052  0.2119  0.5244  0.2192  0.2040  0.0316 -0.4151  0.6189  0.4952
#>  -0.5209  0.4369  0.4236 -0.4185  0.2472 -0.0493  0.2017  0.2300 -0.0616
#> 
#> Columns 19 to 20 -0.1407 -0.1570
#>   0.7811 -0.5050
#>   0.1192 -0.1591
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1053 -0.2570 -0.0264  0.0333 -0.2541 -0.0832  0.3421 -0.5069 -0.0757
#>   0.1353 -0.4522 -0.0671  0.1389  0.0410 -0.1205  0.6665 -0.3621  0.1859
#>   0.0775 -0.2815 -0.1499 -0.1889 -0.0583  0.4660  0.3481  0.0955 -0.2334
#> 
#> Columns 10 to 18 -0.1829 -0.1179  0.1426  0.0283 -0.1242 -0.5858  0.1951  0.4398  0.2916
#>  -0.0940 -0.1406  0.4468  0.2335  0.0280 -0.2375  0.5218  0.3544  0.2055
#>  -0.0314  0.1506 -0.1745  0.2065 -0.1640 -0.3349 -0.0877  0.0007  0.1238
#> 
#> Columns 19 to 20 -0.0556  0.5477
#>   0.2206  0.3775
#>   0.0519  0.1697
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>