RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2687 -0.4477  0.5036  0.6495 -0.4997 -0.0527  0.6919  0.9372  0.6259
#>  -0.6583 -0.6179  0.5020  0.3272  0.2628 -0.4148  0.3177  0.5715 -0.5311
#>   0.8413  0.3005  0.4586 -0.1729 -0.1146 -0.4184 -0.2116  0.6491  0.1596
#> 
#> Columns 10 to 18 -0.2956  0.7433  0.8474  0.5812  0.1965  0.8143  0.9026  0.9086 -0.7776
#>  -0.4945  0.2793 -0.3038  0.1049 -0.1981  0.2243 -0.2652 -0.4303 -0.3677
#>  -0.8979 -0.6189  0.1364 -0.2771  0.0761 -0.0087  0.5866  0.2901 -0.4361
#> 
#> Columns 19 to 20  0.3203 -0.9604
#>   0.3178 -0.3909
#>  -0.0454 -0.0372
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4828  0.2751  0.7671  0.6235  0.2699  0.2893  0.1535  0.7690  0.0382
#>   0.4441 -0.1928  0.6015 -0.0724  0.2396  0.3485  0.4071 -0.1267 -0.2540
#>   0.2692  0.0182  0.3782  0.5672  0.0559  0.3624  0.3385  0.6606 -0.0969
#> 
#> Columns 10 to 18 -0.3416 -0.0927 -0.2818 -0.2705  0.6183  0.5247  0.5362  0.3249 -0.0169
#>  -0.3781 -0.4554 -0.3289 -0.6592  0.2483 -0.4164  0.1267 -0.3239 -0.3380
#>  -0.2184 -0.1859  0.1043  0.0693 -0.0254  0.1894  0.3634 -0.0723 -0.5282
#> 
#> Columns 19 to 20  0.1094 -0.0148
#>  -0.2982 -0.0016
#>   0.4933 -0.6291
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1323  0.6035 -0.0531  0.6191  0.0877  0.2345  0.1049  0.3963  0.0051
#>   0.6620 -0.0473  0.2219  0.1576 -0.2320  0.5241  0.4695  0.5853 -0.0627
#>   0.1728  0.4362  0.5343  0.4158 -0.3370  0.2293  0.3064  0.5245  0.0215
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3781 -0.5353 -0.2560  0.2859  0.2258  0.3461  0.2128 -0.2746 -0.1865
#>   0.6235 -0.6630 -0.2828  0.6924  0.3535 -0.6161  0.5076 -0.0878 -0.3631
#>   0.7065 -0.4504 -0.0703  0.6124  0.3024 -0.2099 -0.0153 -0.4389  0.3658
#> 
#> Columns 10 to 18  0.3674 -0.0695  0.4319  0.4255 -0.0082 -0.0134 -0.2807  0.1293 -0.0784
#>  -0.6755 -0.0982  0.4486  0.6911 -0.1303  0.0743 -0.6025  0.2191 -0.0430
#>  -0.3555 -0.0738  0.0330 -0.0165  0.2915  0.1188 -0.2453 -0.2561  0.2920
#> 
#> Columns 19 to 20 -0.0369  0.3491
#>   0.4072 -0.5225
#>   0.3043 -0.4201
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3361  0.0065  0.3959 -0.0050  0.1585  0.1704  0.2053  0.2412 -0.4050
#>   0.4435  0.5879  0.3304  0.2975  0.0005  0.2143  0.0808  0.4846 -0.0226
#>   0.4270  0.4496  0.3172  0.4537  0.0914  0.3344 -0.0405  0.5452 -0.1019
#> 
#> Columns 10 to 18 -0.2828 -0.2324 -0.3698 -0.0941 -0.0096  0.2810  0.2422 -0.6095 -0.2346
#>  -0.3852 -0.0212 -0.1553  0.1882 -0.1683  0.0104  0.2402 -0.5691  0.0687
#>  -0.5879 -0.2316 -0.3611 -0.1889  0.2677  0.0531  0.0675 -0.3651  0.2738
#> 
#> Columns 19 to 20  0.6215 -0.3107
#>   0.8491  0.1030
#>   0.8461  0.3904
#> [ CPUFloatType{2,3,20} ]
#>