RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0377 -0.2566  0.7807 -0.2172  0.9324 -0.0811  0.1373 -0.4622 -0.8747
#>   0.3252  0.3036  0.1429  0.4040  0.1747 -0.1020 -0.1281 -0.5179 -0.1826
#>   0.5511 -0.6622 -0.3166  0.6750  0.2405  0.4571 -0.6528  0.1645 -0.2248
#> 
#> Columns 10 to 18 -0.1314 -0.2672 -0.5384 -0.5614 -0.6179 -0.9247  0.2017  0.4965 -0.7991
#>   0.2650 -0.0821 -0.1462 -0.6161 -0.1461 -0.9251  0.4926  0.5576 -0.5896
#>  -0.8880 -0.0540  0.6583 -0.4961  0.1318 -0.7003  0.6782  0.3750 -0.0709
#> 
#> Columns 19 to 20 -0.2915  0.6778
#>  -0.2362  0.7635
#>  -0.7516 -0.7706
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5297 -0.4006  0.3951  0.6423 -0.1785 -0.2354  0.3297 -0.1318 -0.6656
#>   0.4213 -0.1735  0.0421  0.4176  0.0988 -0.1065  0.0626  0.1640 -0.0815
#>   0.4221 -0.4354 -0.4189  0.2687  0.3072 -0.4652 -0.0337 -0.2716 -0.1464
#> 
#> Columns 10 to 18  0.0989 -0.2787 -0.2708  0.0325 -0.3909 -0.4170  0.0489  0.0539 -0.5738
#>   0.2309 -0.1023 -0.3117  0.0016 -0.2665 -0.4494 -0.3988 -0.1147 -0.3167
#>   0.0165 -0.4170 -0.4613  0.2623 -0.5155 -0.4079 -0.5348  0.4236  0.5171
#> 
#> Columns 19 to 20  0.0851  0.3440
#>  -0.0920  0.4060
#>  -0.0646 -0.0649
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.3483 -0.5985  0.0714  0.6106  0.3273 -0.1703  0.1830  0.1546 -0.3673
#>  -0.0763 -0.1184  0.4443  0.3341  0.1044  0.3376 -0.0841 -0.0864 -0.4480
#>   0.0522 -0.4772  0.5024  0.5210 -0.0519  0.3913  0.0811 -0.1643 -0.5874
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7782  0.5290  0.2420  0.8392  0.3741 -0.8009 -0.3792 -0.5307  0.1798
#>  -0.0266  0.1164  0.4807  0.1679  0.2227  0.2727 -0.1041 -0.0863 -0.8532
#>  -0.6084 -0.3139 -0.3074  0.2614 -0.4937  0.4690 -0.5139 -0.4079  0.3863
#> 
#> Columns 10 to 18 -0.7991 -0.4087 -0.2625 -0.8587 -0.1486 -0.0907  0.4697 -0.7005 -0.5058
#>   0.4443  0.0811 -0.4984  0.2281  0.3487 -0.5989  0.4623  0.1155 -0.3053
#>  -0.4898 -0.0266  0.5278 -0.4191 -0.0749 -0.0024 -0.2268 -0.4075  0.5409
#> 
#> Columns 19 to 20 -0.1817  0.0340
#>   0.3093 -0.3318
#>   0.6426 -0.4308
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3554 -0.6333  0.5091  0.1830  0.1233 -0.2083  0.0571 -0.1065 -0.2260
#>  -0.1785 -0.3841  0.1587 -0.0267  0.2429  0.0530  0.1925  0.3259  0.0365
#>   0.2572 -0.2762  0.1449  0.5823  0.2217  0.3510 -0.0438 -0.0513 -0.2497
#> 
#> Columns 10 to 18  0.1182 -0.5200 -0.4823  0.2353 -0.0219 -0.7580 -0.3883  0.2841 -0.2847
#>   0.2874  0.1069 -0.0524 -0.0803 -0.3694 -0.6496 -0.5770  0.2189 -0.0763
#>  -0.1075 -0.3094 -0.1697 -0.1638 -0.3259 -0.5657  0.0544  0.5088 -0.6739
#> 
#> Columns 19 to 20 -0.2408  0.5737
#>   0.1398 -0.0037
#>   0.3865  0.6524
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>