RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4041 -0.6079  0.5320 -0.8643 -0.2776  0.2491 -0.2957  0.1326  0.5088
#>  -0.6310  0.3230 -0.0441 -0.7183  0.3832  0.1305 -0.3150 -0.1865  0.5247
#>   0.1830  0.6243  0.1351 -0.0806 -0.2530 -0.1074  0.1755  0.2651  0.3077
#> 
#> Columns 10 to 18 -0.2973  0.2018 -0.1750 -0.2823 -0.6616 -0.6319 -0.6544  0.3596 -0.0050
#>  -0.5898  0.8098  0.3387  0.5436 -0.6345  0.3905 -0.3945 -0.2875  0.5368
#>  -0.5216  0.8565  0.1369  0.6056 -0.6476 -0.3853  0.2539 -0.0759 -0.2852
#> 
#> Columns 19 to 20 -0.9310 -0.7658
#>  -0.1427 -0.6991
#>   0.4274  0.3260
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0291 -0.5084 -0.0839 -0.7487 -0.5402  0.4573  0.1543  0.1905  0.6706
#>   0.1109  0.4227  0.4386 -0.6151  0.2242  0.3873 -0.0731  0.1772  0.2701
#>   0.0712 -0.2207  0.5482 -0.5127  0.1713 -0.0294  0.0430  0.0035  0.2525
#> 
#> Columns 10 to 18  0.2179  0.8190 -0.0949  0.2341  0.1389 -0.0782 -0.3379 -0.2111 -0.2258
#>   0.2624 -0.3981 -0.5358  0.2684  0.5601 -0.1564  0.1182 -0.1160 -0.0020
#>  -0.3327  0.4006  0.1306  0.4493 -0.4185  0.3716  0.2099  0.2060  0.3298
#> 
#> Columns 19 to 20 -0.1322 -0.3444
#>  -0.0396 -0.6087
#>   0.2692 -0.5784
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0396  0.1932  0.3795 -0.5919 -0.0960  0.0227  0.2311 -0.2567  0.6306
#>  -0.1035 -0.0780  0.1739 -0.2390 -0.1490  0.3402  0.2150 -0.3366  0.4363
#>   0.0573 -0.0063  0.2019 -0.6911  0.0436  0.3127  0.1562  0.1409  0.5771
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3990 -0.1244 -0.3920  0.0732  0.6604  0.1913 -0.0532  0.1012 -0.4528
#>  -0.2758  0.6013  0.0499  0.8007 -0.1782 -0.6868 -0.0828 -0.0197 -0.1149
#>  -0.2369  0.5285 -0.5414 -0.0714  0.5946 -0.2822  0.2071 -0.7721 -0.7220
#> 
#> Columns 10 to 18  0.8098 -0.0366  0.1819  0.2124  0.2915  0.3321 -0.2885 -0.4547 -0.7814
#>  -0.1566 -0.3309 -0.1182 -0.1242  0.6716 -0.6082 -0.3269  0.1199 -0.7311
#>  -0.0049 -0.2358  0.5713  0.6567 -0.1848 -0.5074  0.2390  0.0280 -0.3400
#> 
#> Columns 19 to 20 -0.1889 -0.4477
#>  -0.4259 -0.2497
#>  -0.6961 -0.3530
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2906  0.1174  0.1000 -0.4423 -0.1943  0.0281  0.4798 -0.3587  0.5654
#>   0.1187 -0.2872  0.2344 -0.5379 -0.1758  0.0847  0.4037  0.2400  0.7136
#>  -0.0834 -0.1868 -0.1580 -0.3435 -0.4587  0.2911  0.3408 -0.3181  0.5801
#> 
#> Columns 10 to 18  0.0516  0.5292  0.1469  0.1849  0.1084  0.1093  0.0412  0.2499  0.2061
#>  -0.2066  0.6720  0.0990  0.5923 -0.2805 -0.0915 -0.2186  0.4777 -0.1297
#>   0.1669  0.2791  0.1008  0.5555 -0.2700 -0.3153 -0.0824  0.0514 -0.4567
#> 
#> Columns 19 to 20 -0.0942 -0.5616
#>  -0.2170 -0.0982
#>  -0.3382 -0.4198
#> [ CPUFloatType{2,3,20} ]
#>