RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6742  0.2556 -0.3199  0.1142 -0.5693  0.6781 -0.0224  0.2621  0.5746
#>   0.6546  0.4684  0.3814 -0.2833  0.4465  0.6146 -0.5497 -0.5956 -0.0230
#>   0.5866 -0.1263  0.0483  0.2639  0.6731  0.7669  0.5219  0.6074 -0.1219
#> 
#> Columns 10 to 18 -0.2200 -0.7867 -0.6649 -0.2104  0.6104  0.1914  0.7831  0.7697 -0.3085
#>  -0.7173  0.5665 -0.2048  0.2678  0.2632 -0.1421 -0.6138  0.3268 -0.2669
#>  -0.8765 -0.7597  0.2783  0.2507 -0.3471 -0.8639 -0.5192  0.3176  0.5857
#> 
#> Columns 19 to 20 -0.2368  0.4657
#>  -0.7154 -0.2664
#>  -0.4612  0.3307
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3475  0.4304 -0.0873  0.4584  0.4090 -0.0927  0.0748  0.1093 -0.5597
#>  -0.1540 -0.5906 -0.3883 -0.4353  0.3157 -0.0021  0.2118 -0.1667  0.4234
#>   0.4989 -0.3840 -0.6769 -0.6150 -0.5270  0.0140  0.3475 -0.0530  0.5648
#> 
#> Columns 10 to 18  0.3321 -0.0670 -0.1238  0.4559  0.2615 -0.2090 -0.3598  0.3164  0.2441
#>  -0.3207  0.0757 -0.6479  0.3037  0.0933 -0.3857  0.2859  0.2246 -0.4824
#>   0.2730 -0.7531  0.0467  0.1405  0.1624 -0.7932  0.3616  0.5495 -0.1700
#> 
#> Columns 19 to 20 -0.4864 -0.1104
#>  -0.3677  0.3429
#>  -0.5615  0.4974
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4063 -0.4911 -0.2177 -0.2842 -0.0390  0.5646  0.2498  0.1631 -0.0862
#>   0.3794  0.5999  0.3541  0.2216 -0.0542  0.0654 -0.1534  0.1803 -0.2483
#>   0.2789  0.6703 -0.1528  0.0385  0.2029  0.0796 -0.1976  0.1821  0.1311
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6029  0.4150  0.2196  0.3697 -0.0238  0.4424  0.5882 -0.5121  0.6942
#>  -0.4032 -0.6584  0.3615  0.2037 -0.6263  0.0652 -0.3856  0.0413  0.5756
#>  -0.3440 -0.2449  0.3821 -0.1451 -0.3360  0.4328  0.5131 -0.0124  0.4596
#> 
#> Columns 10 to 18  0.1801 -0.8158  0.6796 -0.3487 -0.0430 -0.2465  0.3077  0.4187  0.1662
#>  -0.0228  0.2641 -0.0463 -0.5728 -0.1061  0.0488 -0.6136 -0.5340  0.0737
#>   0.4532 -0.5304 -0.1179 -0.1546  0.1628 -0.1487 -0.0016  0.0949  0.3003
#> 
#> Columns 19 to 20  0.4417 -0.5388
#>  -0.3167 -0.2277
#>   0.2817  0.5181
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0609  0.1176 -0.6241  0.3022 -0.5453 -0.3246  0.0606  0.4984  0.3136
#>   0.1968 -0.1635 -0.3244 -0.1955 -0.2538  0.3740 -0.0125  0.5638 -0.0786
#>   0.0522  0.2553 -0.3174  0.0056 -0.4417 -0.2641  0.0717  0.5171  0.0534
#> 
#> Columns 10 to 18 -0.3759  0.0853  0.4720 -0.2116  0.3987 -0.6387  0.0324  0.2173 -0.0782
#>  -0.1286 -0.4678  0.2524  0.2421  0.0554 -0.5543  0.3380  0.4284 -0.2649
#>  -0.2088 -0.3520 -0.1102 -0.1728  0.4843 -0.4790  0.3425  0.1760 -0.2649
#> 
#> Columns 19 to 20 -0.2956 -0.1820
#>  -0.5811  0.4048
#>  -0.2604 -0.0285
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>