RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1833  0.6560  0.1257 -0.0904  0.4891 -0.3603 -0.1009  0.0680 -0.6214
#>  -0.8298  0.8248 -0.0496 -0.6105 -0.6182  0.8994 -0.5695 -0.5256 -0.7462
#>  -0.1099 -0.3413  0.3327 -0.5667 -0.0130 -0.1269 -0.8478 -0.6131 -0.3753
#> 
#> Columns 10 to 18 -0.3383 -0.0481 -0.2184 -0.5172 -0.0207  0.4452 -0.1817 -0.9325 -0.2829
#>   0.0268  0.3389  0.1406  0.7982 -0.4113  0.5181 -0.2747 -0.6402  0.1708
#>  -0.6067  0.3039 -0.1021 -0.3474  0.0148 -0.3818  0.7597 -0.8111  0.6746
#> 
#> Columns 19 to 20  0.1399  0.0352
#>  -0.8338  0.1303
#>   0.7177 -0.0482
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0788  0.4503  0.3966 -0.2173 -0.0970  0.3462  0.4994 -0.0375  0.6102
#>   0.3270  0.6634  0.3290 -0.2321 -0.7195  0.0345 -0.1575  0.4668 -0.4615
#>  -0.0157  0.2505  0.6332 -0.4091  0.5466  0.4530 -0.1600 -0.1290  0.1821
#> 
#> Columns 10 to 18  0.2054 -0.0101 -0.3663 -0.0748 -0.3918  0.2464 -0.2180 -0.0456 -0.2173
#>  -0.4908  0.6634 -0.4525 -0.3158 -0.5318 -0.3280 -0.4755 -0.0988 -0.4643
#>  -0.1393 -0.0828 -0.8413  0.0160 -0.6847  0.5600 -0.4339 -0.5882 -0.0151
#> 
#> Columns 19 to 20 -0.1952 -0.5199
#>   0.6953 -0.5109
#>  -0.1261 -0.3081
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0955  0.0990  0.1709 -0.4023 -0.3945  0.4151 -0.0302  0.4181  0.1048
#>  -0.2089  0.3208  0.3037 -0.4527 -0.3052  0.5980 -0.0021 -0.3411  0.3968
#>   0.1317  0.2469  0.3336 -0.4281 -0.0248  0.0547  0.1274  0.6921  0.2598
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0697  0.0270  0.2264 -0.3314 -0.0338 -0.0901  0.9252  0.4531  0.4889
#>  -0.1770  0.2884  0.3648  0.3197 -0.0109 -0.6168  0.7644  0.4215  0.4784
#>   0.3784  0.4804  0.3072  0.5074 -0.0618 -0.4803  0.3960  0.4399  0.0146
#> 
#> Columns 10 to 18 -0.2134 -0.2791 -0.6203 -0.5011 -0.5898  0.0160  0.5624 -0.2182 -0.2067
#>   0.2516 -0.7565  0.2071  0.4511  0.1258 -0.4953  0.1788  0.4434 -0.6428
#>   0.3790 -0.5221  0.3225  0.1566  0.3454 -0.5043 -0.5658  0.4697  0.0110
#> 
#> Columns 19 to 20 -0.0515 -0.7048
#>   0.0593 -0.5197
#>  -0.0510 -0.1628
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3658 -0.2102  0.2837 -0.5025  0.2696  0.0981  0.4119  0.4663  0.0843
#>   0.1705 -0.3611  0.2420 -0.5864  0.0602 -0.0813 -0.3575  0.1974  0.1472
#>  -0.6387  0.1682  0.3260 -0.1248 -0.1548  0.2370 -0.0247  0.0886  0.2233
#> 
#> Columns 10 to 18 -0.2401  0.1487 -0.4070  0.2028 -0.3987  0.1912 -0.0948 -0.5403 -0.4252
#>  -0.2676  0.0796 -0.5072  0.1463 -0.2901  0.1172 -0.0529 -0.7666 -0.1760
#>   0.1799 -0.2038 -0.4256 -0.3471 -0.1832  0.3382 -0.1427 -0.4738  0.1017
#> 
#> Columns 19 to 20 -0.1083 -0.3116
#>   0.2095 -0.1415
#>   0.4518  0.0721
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>