RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6614  0.5800 -0.2110  0.0198  0.0520  0.6427 -0.1871  0.5312 -0.4566
#>   0.7076  0.2240  0.5183 -0.5605  0.5341  0.1096 -0.1854  0.7758  0.9505
#>   0.1839  0.3832  0.3556  0.0282 -0.8587 -0.4531  0.2706 -0.1431  0.8904
#> 
#> Columns 10 to 18  0.7514  0.6862 -0.5468  0.2683 -0.5484  0.0991  0.3324 -0.5348 -0.7286
#>  -0.9333  0.2488 -0.6675  0.8323  0.7379 -0.3567 -0.1124 -0.4047 -0.1688
#>   0.9320 -0.0676  0.9248 -0.2598 -0.4656 -0.7853 -0.6703  0.4639 -0.1490
#> 
#> Columns 19 to 20  0.7163 -0.6247
#>   0.6400  0.1039
#>  -0.8210  0.6889
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3436  0.3477 -0.3458  0.3207  0.1049 -0.7244 -0.1529  0.4866 -0.1057
#>   0.4560  0.4785 -0.0583  0.5022 -0.0335 -0.0931  0.3857  0.3462  0.2300
#>   0.5594  0.4112 -0.1958  0.3426  0.3557 -0.2781  0.0168 -0.5522  0.6289
#> 
#> Columns 10 to 18  0.0133  0.8164 -0.0769  0.1406 -0.0960 -0.2362  0.1074  0.3133  0.4502
#>  -0.3362 -0.5406  0.2617  0.1781 -0.0561 -0.1906 -0.5599  0.0760  0.4846
#>  -0.2341  0.3004 -0.3100 -0.4520 -0.1888 -0.3720 -0.4914 -0.0393  0.6264
#> 
#> Columns 19 to 20  0.4272 -0.6527
#>   0.6782 -0.2300
#>   0.4952  0.4094
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1376  0.5023 -0.1298  0.0863  0.0515  0.1877 -0.2470  0.4897  0.3602
#>   0.0378 -0.1608  0.0947  0.4092 -0.5453 -0.2727 -0.3904  0.1253  0.1021
#>   0.4752 -0.0793  0.3417  0.7543  0.0170 -0.3323 -0.6362 -0.1338 -0.0704
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1902  0.2711 -0.2633  0.1763 -0.6384  0.0324  0.1745 -0.4699 -0.1281
#>  -0.0964  0.2175  0.1496 -0.4801 -0.2767  0.1351  0.1647  0.1430 -0.4375
#>  -0.5673 -0.4420  0.1037 -0.1998  0.5650 -0.4852  0.6845 -0.2764 -0.6209
#> 
#> Columns 10 to 18  0.5110 -0.7814  0.2304 -0.2415  0.1657  0.1336 -0.3273  0.1822  0.6765
#>   0.2783 -0.2454 -0.4368 -0.2611 -0.4027  0.0157  0.0360 -0.2148  0.5617
#>   0.2603  0.1933 -0.2942 -0.8786  0.1780  0.0263 -0.5088  0.4326 -0.6948
#> 
#> Columns 19 to 20 -0.2077 -0.8455
#>   0.1151 -0.5150
#>   0.9357  0.8474
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4336  0.0882 -0.3104  0.1999 -0.3144 -0.4189 -0.4412  0.2131 -0.3355
#>   0.1251 -0.0096  0.0306  0.6179 -0.5781 -0.4510 -0.3960  0.1264 -0.0356
#>   0.1806  0.4130 -0.0863  0.8623  0.0099  0.4071 -0.6038  0.2020 -0.0098
#> 
#> Columns 10 to 18  0.3525  0.2429  0.0182  0.4614 -0.5048 -0.2627  0.0883 -0.3914 -0.5055
#>  -0.2276  0.2887  0.1210  0.3815 -0.0093 -0.1153  0.0141 -0.0449 -0.2600
#>  -0.1270 -0.1645 -0.3460 -0.1325 -0.5359 -0.0510 -0.1582 -0.1119 -0.0790
#> 
#> Columns 19 to 20  0.4875 -0.1788
#>   0.5460  0.0530
#>   0.1332  0.1056
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>