RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4366  0.5784 -0.0080  0.8192  0.1028 -0.2242 -0.7217 -0.1797 -0.2464
#>   0.0834  0.7164  0.8921 -0.5294 -0.5007  0.5961 -0.8081  0.3294  0.6644
#>   0.8122  0.6071 -0.6713  0.8243  0.5760 -0.6334 -0.5883  0.5789 -0.3453
#> 
#> Columns 10 to 18  0.0093 -0.8995 -0.2572  0.1226 -0.4624  0.2280 -0.6314  0.4888 -0.1991
#>   0.5026 -0.6166  0.0361  0.3980 -0.8740 -0.1536 -0.3048  0.2346  0.3281
#>  -0.0601 -0.2540  0.1590 -0.2710  0.2365  0.1195 -0.2861 -0.8235 -0.0771
#> 
#> Columns 19 to 20  0.7361  0.2263
#>  -0.6490  0.8519
#>   0.2851  0.8556
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4941  0.1146 -0.2825  0.6518 -0.6052  0.6934  0.0391  0.3548  0.2010
#>   0.4311  0.5497 -0.3692  0.7892 -0.5413  0.6436 -0.4045  0.7392 -0.5336
#>   0.4103 -0.3217 -0.2047 -0.4835  0.3376  0.2355 -0.4194 -0.5338  0.0785
#> 
#> Columns 10 to 18  0.6596 -0.4048 -0.5137 -0.2415 -0.2093  0.5214  0.1139 -0.6080  0.5843
#>   0.6760 -0.2862  0.4041 -0.0212 -0.0842  0.0429 -0.2672 -0.6396 -0.3478
#>   0.1971 -0.0932 -0.0413 -0.2115 -0.0695  0.4326  0.7655 -0.1270  0.2846
#> 
#> Columns 19 to 20 -0.3934  0.0042
#>   0.0910 -0.0461
#>  -0.3107 -0.0836
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0846  0.2102  0.1906  0.3355  0.5423  0.1918  0.2006  0.7216  0.1118
#>   0.5639  0.3293 -0.1475  0.4413  0.2220  0.1953 -0.0873  0.2573 -0.2926
#>   0.1754 -0.4530  0.2322 -0.0763  0.3614  0.1058  0.0711  0.0879 -0.2773
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4818 -0.0543  0.2963 -0.7118 -0.5325 -0.0297  0.5198 -0.0710 -0.4479
#>  -0.5790  0.4157 -0.6559 -0.2132 -0.1187 -0.4835  0.2906 -0.1246 -0.6056
#>  -0.1526 -0.6523  0.3932 -0.2259 -0.2823 -0.0809  0.4502 -0.6554 -0.3895
#> 
#> Columns 10 to 18 -0.2617  0.5791 -0.2805 -0.5405 -0.0318 -0.5498  0.3082  0.2240 -0.2676
#>   0.3679  0.7764 -0.1211  0.3535 -0.7098  0.3028  0.0330  0.1883  0.0750
#>  -0.0629 -0.0184 -0.6096  0.5245  0.5356 -0.5811 -0.4451 -0.8695  0.2877
#> 
#> Columns 19 to 20 -0.2020  0.4934
#>  -0.5729 -0.0410
#>  -0.0574 -0.3621
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3021  0.5985  0.2181  0.6720  0.0471  0.3317 -0.0197  0.5906  0.1519
#>   0.2071 -0.1318  0.3319  0.1347  0.1390  0.3760 -0.0395  0.1769 -0.0455
#>   0.2543  0.5409  0.3945  0.3349 -0.2895 -0.0139 -0.3865 -0.0102  0.2910
#> 
#> Columns 10 to 18  0.5123 -0.0976 -0.0923  0.2748 -0.1860  0.1626 -0.1059  0.0791 -0.2807
#>  -0.0301 -0.1773 -0.1520  0.0009  0.2523 -0.2179  0.4231 -0.2656  0.2717
#>   0.0501 -0.6798 -0.5055 -0.0383 -0.0579  0.5293 -0.4528  0.1573 -0.2576
#> 
#> Columns 19 to 20 -0.1432  0.1724
#>  -0.1624 -0.4178
#>  -0.1964  0.4079
#> [ CPUFloatType{2,3,20} ]
#>