RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0373  0.1002 -0.6331  0.0019 -0.7810 -0.6533  0.6248  0.7642 -0.8411
#>   0.4468 -0.6737  0.3562  0.2438  0.5601 -0.3357  0.8112  0.7616  0.2169
#>   0.8515  0.5209  0.3503  0.4046 -0.1754  0.9052  0.4024  0.0690 -0.1684
#> 
#> Columns 10 to 18 -0.1679  0.3646 -0.0640  0.1649  0.0895  0.4527 -0.4991 -0.0359 -0.1352
#>   0.2582  0.4790  0.0286 -0.0144 -0.3717  0.6759  0.2621 -0.7898  0.3493
#>   0.3407 -0.1168  0.6369 -0.1935  0.7522  0.5462  0.8624  0.1227 -0.3897
#> 
#> Columns 19 to 20  0.5355 -0.1836
#>  -0.1129 -0.3365
#>  -0.4727 -0.0361
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1151 -0.4004 -0.1674 -0.4344 -0.4186 -0.0721 -0.0183  0.3998  0.4594
#>   0.0424 -0.0989  0.1399  0.4555 -0.1319  0.5505  0.2531  0.6271 -0.4225
#>  -0.3698 -0.7057 -0.0813  0.2695 -0.0908  0.0388  0.2445  0.4737 -0.5587
#> 
#> Columns 10 to 18 -0.4170 -0.0949  0.1086  0.1762  0.4246 -0.5781 -0.7064  0.2655 -0.1876
#>   0.3381  0.5450 -0.1032  0.0384  0.5162 -0.3959 -0.0360  0.2923  0.0729
#>   0.5815  0.6084 -0.2009 -0.6353 -0.6300  0.3252 -0.6386 -0.0444 -0.7979
#> 
#> Columns 19 to 20  0.2154 -0.2223
#>   0.1960  0.2044
#>  -0.1919 -0.5982
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0378 -0.4283  0.0493  0.4490 -0.3420 -0.1608 -0.1161  0.5243  0.3873
#>  -0.3010 -0.5004 -0.0474  0.2791 -0.1751 -0.1278  0.1593  0.3835 -0.0979
#>  -0.3547 -0.2648  0.1191  0.5014 -0.2936 -0.2836 -0.3307  0.0130  0.4512
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3580  0.5311 -0.4713 -0.1466  0.1305  0.2479  0.0190 -0.1600  0.3486
#>   0.5636  0.1138 -0.8248 -0.3722  0.1284  0.1649  0.4943 -0.6362  0.0072
#>   0.7349 -0.8398 -0.6293 -0.5774  0.3118 -0.0989  0.8429 -0.8911 -0.6207
#> 
#> Columns 10 to 18 -0.3226 -0.4721  0.3073 -0.2832  0.0929  0.2153 -0.1379 -0.1831  0.6711
#>  -0.5586 -0.8003  0.6995 -0.0536  0.2629 -0.3144 -0.2597 -0.5794  0.4369
#>  -0.1007 -0.1237 -0.1317  0.0725  0.2162 -0.7631  0.7467  0.2450  0.5756
#> 
#> Columns 19 to 20  0.1317 -0.5698
#>  -0.0252 -0.0112
#>   0.2934  0.1933
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1716 -0.0306  0.5763  0.2827 -0.3714 -0.1289 -0.2086  0.2698 -0.1772
#>  -0.2971 -0.4876  0.5636  0.3536 -0.3485  0.0758 -0.3712 -0.1395  0.1480
#>   0.0764 -0.5475 -0.2441  0.5279 -0.5368  0.5551 -0.1759 -0.1610 -0.0678
#> 
#> Columns 10 to 18 -0.3579 -0.0163 -0.1640 -0.4321 -0.0490  0.0713 -0.4823 -0.0545 -0.1881
#>  -0.0073  0.0740 -0.5029 -0.4496 -0.3268  0.2822 -0.5217  0.0506 -0.1209
#>   0.1238  0.0028 -0.0101 -0.6390 -0.0803  0.3344 -0.2016  0.3780 -0.2712
#> 
#> Columns 19 to 20  0.0442 -0.2130
#>   0.5165 -0.2549
#>   0.4474  0.2073
#> [ CPUFloatType{2,3,20} ]
#>