RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5514  0.5645 -0.3749  0.6692  0.3751 -0.3874  0.0458  0.1654  0.5259
#>   0.5033 -0.8981 -0.0589 -0.3201  0.9249 -0.0043 -0.5251 -0.6073 -0.6835
#>  -0.3066  0.2524  0.3933  0.2481  0.0238 -0.7130 -0.8112 -0.7877  0.4785
#> 
#> Columns 10 to 18 -0.1690  0.0929  0.0119 -0.3284 -0.0537 -0.6850  0.6911 -0.8153  0.3462
#>   0.1607  0.5816  0.8524  0.8186  0.0655  0.1884  0.2877  0.8448  0.2269
#>   0.7359  0.0258  0.8450  0.8852 -0.8648 -0.4320  0.5719  0.5567 -0.3042
#> 
#> Columns 19 to 20 -0.0647  0.2102
#>   0.6255  0.8065
#>  -0.5553 -0.0865
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0823  0.1956  0.3399 -0.0027 -0.0193  0.0118 -0.4200 -0.7319  0.2636
#>   0.4331 -0.0521  0.1521 -0.0513 -0.0528  0.1482  0.2060 -0.2593  0.0072
#>   0.6003 -0.2877  0.6790  0.4185  0.3246 -0.2564  0.3051 -0.2478 -0.1781
#> 
#> Columns 10 to 18 -0.0484 -0.0643  0.7804  0.5990  0.1554 -0.3552  0.0658  0.2043 -0.5842
#>  -0.4697  0.4005  0.8115 -0.3201 -0.3212 -0.5424 -0.4583 -0.3637 -0.1312
#>  -0.3351 -0.0439  0.6658 -0.2644  0.6246 -0.1347 -0.3874 -0.6032  0.5381
#> 
#> Columns 19 to 20  0.5123 -0.0069
#>  -0.4736  0.1902
#>  -0.0305 -0.3825
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0144 -0.4054  0.0401  0.0927  0.4168 -0.5578  0.3454 -0.1520  0.4117
#>  -0.1051  0.4831  0.2920 -0.2864  0.1139 -0.3021  0.1306 -0.2153  0.4897
#>  -0.6075  0.0731 -0.4722 -0.0418 -0.0643 -0.0889 -0.0107  0.3345  0.3327
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3222  0.4431  0.2514 -0.6191  0.8375  0.2048  0.3463 -0.7673  0.5565
#>  -0.3454  0.4998  0.4778 -0.7224  0.1033  0.4531 -0.6596  0.6656 -0.7241
#>  -0.1229  0.5043  0.8105  0.6697 -0.2965 -0.3624  0.5852 -0.1389 -0.4774
#> 
#> Columns 10 to 18  0.5812  0.0473  0.0103  0.8091 -0.4483 -0.7437  0.7370 -0.2910 -0.8645
#>   0.1691 -0.1044  0.1619 -0.1715 -0.1781  0.0575  0.5878  0.1842 -0.1746
#>   0.4895 -0.1825  0.3943  0.8928 -0.3062 -0.7677 -0.6904 -0.8840  0.9011
#> 
#> Columns 19 to 20  0.0769  0.5375
#>   0.0140  0.7056
#>  -0.4172 -0.7692
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0644  0.4009  0.3598 -0.2633  0.1702 -0.0889  0.0177 -0.3401  0.4791
#>  -0.1671  0.0801  0.1494 -0.0856 -0.2070  0.2051 -0.7236 -0.3435 -0.1826
#>  -0.0689  0.3211 -0.1065 -0.3428 -0.4546  0.1555 -0.6578 -0.3243  0.1029
#> 
#> Columns 10 to 18 -0.0989  0.0019  0.7467  0.3433  0.0335 -0.3299 -0.2249 -0.1285 -0.2310
#>  -0.0373  0.0818  0.3825  0.0995 -0.2583 -0.4129 -0.0687 -0.5520  0.0108
#>   0.1847  0.1603  0.4556 -0.3041  0.0768 -0.5160 -0.4040 -0.4425 -0.4301
#> 
#> Columns 19 to 20  0.3512  0.1795
#>  -0.0052  0.1372
#>   0.3553  0.0013
#> [ CPUFloatType{2,3,20} ]
#>