RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2182 -0.4666  0.0822  0.6045  0.6101  0.1501  0.2770  0.5292 -0.8082
#>   0.0048  0.3354 -0.8594 -0.2090 -0.3306 -0.7107  0.4221  0.4244  0.2569
#>  -0.6844 -0.2187 -0.8386 -0.1192 -0.3578 -0.3256  0.5279  0.4233  0.8748
#> 
#> Columns 10 to 18  0.4740 -0.2778 -0.4865 -0.8878  0.8593 -0.3555 -0.6097  0.5003 -0.1595
#>  -0.7558 -0.3329  0.1491 -0.6706 -0.0437 -0.0090 -0.3052 -0.0232 -0.6322
#>  -0.1771 -0.4217  0.4726 -0.8009  0.1755  0.7799  0.1764  0.6219 -0.2131
#> 
#> Columns 19 to 20  0.2728 -0.4146
#>   0.4222  0.5648
#>  -0.3236  0.6146
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1164 -0.6558 -0.3480  0.7560 -0.5921  0.2957  0.2270 -0.0510  0.0017
#>  -0.0603 -0.0848 -0.5830  0.4153 -0.6382  0.1508  0.2362  0.2537  0.2119
#>  -0.3032 -0.3676 -0.0525  0.0041 -0.0537  0.3242  0.3482  0.3666  0.3958
#> 
#> Columns 10 to 18 -0.2303 -0.4819  0.2042 -0.3342 -0.1390  0.2344  0.4098  0.7577  0.3883
#>  -0.4919 -0.4129  0.0021 -0.1234  0.2284 -0.1223  0.0799  0.5683 -0.0201
#>  -0.1776  0.1268 -0.0011 -0.0891 -0.2840 -0.0959 -0.1390  0.5945  0.3489
#> 
#> Columns 19 to 20 -0.4306 -0.2669
#>  -0.0334  0.2546
#>  -0.0458  0.3423
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1490 -0.3914 -0.4614  0.4271  0.0102  0.0245 -0.0143 -0.3756  0.5387
#>  -0.2782 -0.5626 -0.2191  0.3218 -0.4369  0.2705  0.1127 -0.1594  0.0422
#>  -0.5438 -0.4132 -0.3866  0.2035 -0.1068  0.3859 -0.2054 -0.2298  0.4769
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4345  0.2606  0.3526  0.5677 -0.7161 -0.2888 -0.0913  0.8311  0.1686
#>  -0.2452  0.9037  0.4792  0.3115  0.5448 -0.5077 -0.6213 -0.3342 -0.7002
#>  -0.6805  0.7806  0.4984  0.4113  0.2254  0.1401 -0.7127  0.5320 -0.2122
#> 
#> Columns 10 to 18 -0.3033  0.8423  0.1852  0.3486  0.0461 -0.1428  0.2833 -0.0944  0.1417
#>  -0.0300  0.1913 -0.3439  0.1179  0.4539  0.5515 -0.1082 -0.3554  0.0580
#>   0.4100  0.5087 -0.1901  0.2387 -0.0731 -0.2838  0.5188 -0.6386 -0.1538
#> 
#> Columns 19 to 20 -0.0619  0.3214
#>  -0.4691 -0.3601
#>   0.3255  0.5226
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0560 -0.4193 -0.1186  0.3852 -0.0875  0.5047 -0.0656 -0.1090  0.2897
#>   0.0685 -0.5749 -0.2357  0.2567 -0.2050  0.6067  0.2023  0.1183 -0.3649
#>  -0.3228 -0.5377 -0.4851  0.5706 -0.3442  0.5482 -0.1384 -0.0990  0.2859
#> 
#> Columns 10 to 18 -0.8278 -0.4703 -0.3535  0.0202 -0.2313 -0.4616 -0.2472  0.3138 -0.4815
#>   0.2569 -0.1156  0.0495 -0.1818 -0.0614 -0.3880 -0.4222  0.3443  0.1341
#>  -0.5380 -0.4805 -0.2283  0.0392 -0.2218 -0.4536  0.2956  0.6263 -0.2344
#> 
#> Columns 19 to 20  0.2523  0.1412
#>   0.2858  0.1571
#>   0.4027  0.2483
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>