RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5501  0.5635 -0.3553 -0.6537  0.5293  0.8198 -0.7474  0.4945  0.4636
#>  -0.6285 -0.5403  0.6875  0.6013  0.0741  0.1435 -0.3812 -0.3252  0.8580
#>   0.1813 -0.8171  0.0672 -0.1077 -0.4799  0.6041  0.6916  0.5862 -0.4076
#> 
#> Columns 10 to 18  0.6839  0.4395  0.6885  0.1652 -0.6109  0.2467  0.3743  0.0032  0.5608
#>   0.6584 -0.6390  0.6794  0.5947  0.1192  0.0954  0.6923 -0.2665  0.6602
#>   0.9049  0.2189  0.2200  0.6244  0.5116 -0.9232 -0.7063 -0.7814 -0.3636
#> 
#> Columns 19 to 20 -0.2320  0.4104
#>   0.2093 -0.6555
#>   0.7413  0.1361
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5878 -0.3544 -0.0121 -0.0907  0.0313 -0.0509  0.1906  0.1384  0.2382
#>  -0.0711 -0.1650  0.5208  0.2012  0.1474  0.0006 -0.7109  0.1717  0.7293
#>   0.0636 -0.3923 -0.5041 -0.0325  0.2203 -0.4049 -0.5177 -0.2532  0.6659
#> 
#> Columns 10 to 18 -0.1624  0.0869  0.0994 -0.6315 -0.1345 -0.5366 -0.1483  0.4541  0.0782
#>   0.5274 -0.0360  0.6822  0.3201 -0.1955 -0.1293 -0.3517 -0.1095  0.4225
#>   0.0644 -0.5084  0.3857  0.0833 -0.5659  0.3354 -0.0670 -0.6695  0.1844
#> 
#> Columns 19 to 20 -0.1533  0.4769
#>  -0.5641  0.3666
#>   0.3153 -0.2123
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5930 -0.2719 -0.0919  0.0169 -0.0214 -0.0166  0.1897  0.2000  0.1967
#>   0.1147 -0.0898 -0.2142  0.2421 -0.3135 -0.0589 -0.1671  0.0399  0.5889
#>  -0.0909 -0.0422 -0.4449 -0.4583  0.0859  0.4687 -0.0130  0.2905 -0.2470
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1699 -0.0090  0.2468  0.2623 -0.5376  0.1850 -0.6395 -0.4066  0.6195
#>  -0.1032  0.1002 -0.2362  0.3955  0.0961 -0.2371  0.3904 -0.4787  0.3214
#>  -0.3662  0.1250  0.5315 -0.2120 -0.2916 -0.2393  0.1219 -0.5774  0.7791
#> 
#> Columns 10 to 18 -0.1444 -0.5152 -0.0149 -0.6742  0.2266  0.2252 -0.0373 -0.8838  0.5789
#>   0.5836  0.1184 -0.2856 -0.6383 -0.2191  0.0692 -0.2795  0.1347  0.4028
#>   0.5148  0.6385  0.0786 -0.7139 -0.3213  0.0655 -0.2651 -0.3243  0.4697
#> 
#> Columns 19 to 20  0.0302  0.1147
#>   0.5385  0.6137
#>   0.1091  0.0247
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1077 -0.3072  0.0605 -0.0738 -0.1108 -0.3108 -0.3211  0.2128  0.0637
#>  -0.4582 -0.3945 -0.3493 -0.0194  0.4062 -0.1334 -0.2800  0.5833  0.2202
#>  -0.2967 -0.1904 -0.5358 -0.3591  0.4390 -0.5226  0.0202  0.6168 -0.2294
#> 
#> Columns 10 to 18  0.2636  0.0017  0.2574 -0.1542 -0.0424 -0.0200  0.1070  0.1958  0.0513
#>   0.0062 -0.2492  0.0167 -0.0901 -0.1512  0.1503  0.1113  0.0514 -0.0037
#>   0.1156 -0.1826  0.0375 -0.0138 -0.0724  0.4764  0.1518  0.5159 -0.0632
#> 
#> Columns 19 to 20 -0.4255  0.2016
#>  -0.3730  0.0412
#>  -0.5335  0.0865
#> [ CPUFloatType{2,3,20} ]
#>