RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1139 -0.0374 -0.8699  0.1981 -0.4858  0.0816 -0.7914 -0.1448 -0.4141
#>  -0.2952 -0.6266 -0.4393  0.3406  0.1555 -0.6662 -0.8674  0.2755  0.2647
#>  -0.4995  0.6348 -0.5048 -0.9126 -0.3617  0.8597 -0.4468  0.4653 -0.3341
#> 
#> Columns 10 to 18 -0.5143 -0.5790 -0.7068 -0.8624 -0.3752 -0.9103 -0.3916  0.7050 -0.0650
#>   0.1331  0.3585 -0.3873  0.1451  0.8184 -0.2981  0.7079 -0.1282  0.1850
#>  -0.2917 -0.6859  0.5325 -0.0649  0.1048 -0.1288  0.8252 -0.1055 -0.3046
#> 
#> Columns 19 to 20  0.7965  0.2105
#>  -0.1722 -0.7687
#>  -0.8399 -0.1448
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0286  0.4037 -0.4488  0.6682 -0.4789 -0.5752 -0.2407  0.2392  0.2331
#>   0.1586 -0.4621  0.3114  0.1684 -0.2428  0.5763  0.3074  0.6288 -0.5176
#>   0.1384  0.2744 -0.2710  0.1432 -0.4072 -0.3363 -0.2498  0.1667  0.3060
#> 
#> Columns 10 to 18 -0.1956 -0.0217 -0.5542  0.6839  0.5964 -0.4499 -0.0217 -0.4132  0.1234
#>  -0.4186  0.2304 -0.5461 -0.2943 -0.2724 -0.4728  0.3897  0.1859  0.6147
#>   0.2220 -0.4216 -0.4553 -0.0803  0.2739 -0.3961 -0.4335  0.5363  0.1241
#> 
#> Columns 19 to 20  0.8442  0.2680
#>   0.3015  0.3750
#>   0.2162  0.5263
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0643  0.2163 -0.6463 -0.1692  0.2111  0.1883 -0.0655  0.3924 -0.2593
#>   0.1980  0.3056 -0.3408 -0.3856 -0.2948 -0.2301 -0.5316  0.2743  0.0425
#>   0.1069 -0.2779 -0.6762 -0.0379 -0.1617 -0.6576 -0.5346  0.6562  0.6149
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0505  0.0025 -0.0056 -0.6749 -0.2656  0.3406 -0.6426 -0.0122  0.2062
#>  -0.6411 -0.8289 -0.2801 -0.4265  0.5385 -0.8922 -0.8250 -0.6133  0.5507
#>  -0.0915  0.1267  0.2429  0.1249  0.1366 -0.1224  0.0994 -0.6059  0.2512
#> 
#> Columns 10 to 18  0.0855  0.3084 -0.2918 -0.5998  0.5625 -0.0377 -0.1179  0.0082  0.3540
#>  -0.2737 -0.4787  0.1477  0.3759  0.5983  0.5078 -0.5774  0.2238  0.2213
#>   0.2735 -0.3156 -0.1265  0.1086  0.4599 -0.7744 -0.0175 -0.4649  0.3988
#> 
#> Columns 19 to 20  0.0151  0.3063
#>  -0.5036  0.7729
#>   0.3660 -0.1472
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1214  0.2485 -0.6725 -0.0934  0.1811 -0.5381 -0.2191  0.5052  0.0967
#>   0.0101 -0.3666 -0.5331  0.1002 -0.1833 -0.2326 -0.1169  0.3389  0.6608
#>   0.1184  0.2270 -0.0476 -0.0690 -0.3195 -0.0151 -0.1080  0.2877 -0.3110
#> 
#> Columns 10 to 18 -0.0182 -0.1059 -0.4959  0.0601  0.1788  0.1770  0.2545 -0.1617  0.0341
#>  -0.4228 -0.5245 -0.5912 -0.4599  0.3470 -0.2404  0.1210 -0.0185  0.1092
#>  -0.1583 -0.2480 -0.5081 -0.2151  0.2153 -0.1600  0.0176 -0.3467  0.3699
#> 
#> Columns 19 to 20 -0.1266 -0.2928
#>   0.3619  0.3422
#>   0.1418  0.0821
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>