RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4151  0.5544 -0.3372 -0.5162 -0.7199  0.3701 -0.0377 -0.3255  0.7345
#>  -0.4822  0.5009 -0.4706 -0.5086 -0.1751 -0.1331 -0.5769  0.1420  0.4743
#>  -0.5365  0.5203 -0.4053 -0.1174 -0.6265  0.6999 -0.0134 -0.0400 -0.5242
#> 
#> Columns 10 to 18 -0.5962  0.4256  0.9351  0.4741  0.1131 -0.1164 -0.1743 -0.6670 -0.3001
#>  -0.0886  0.3078 -0.8030  0.4994 -0.0215 -0.8993  0.2812 -0.6938 -0.8218
#>   0.6437  0.8481 -0.1849  0.7082 -0.5708  0.1974 -0.5187  0.2992 -0.1995
#> 
#> Columns 19 to 20 -0.6211 -0.7527
#>  -0.9871 -0.6786
#>  -0.6053 -0.4999
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.7339  0.0262  0.2451  0.7772 -0.3094 -0.1288  0.7075 -0.0073 -0.0776
#>  -0.1862  0.1401 -0.3950 -0.0220 -0.7383  0.1007 -0.2562 -0.4306 -0.3075
#>  -0.0947  0.0039 -0.2057  0.3889 -0.3187 -0.2029 -0.3219 -0.4683 -0.5678
#> 
#> Columns 10 to 18 -0.3649 -0.1937 -0.2845  0.3953  0.2210 -0.7414  0.5216  0.5884  0.5433
#>  -0.0196 -0.1520 -0.2901  0.1603 -0.3785 -0.4082 -0.0262  0.3492  0.2054
#>  -0.1102 -0.6930 -0.5360  0.6392  0.4047 -0.6335  0.0753  0.7952 -0.3014
#> 
#> Columns 19 to 20  0.5072  0.1192
#>  -0.2842  0.2489
#>   0.0722  0.3707
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3209 -0.0198  0.5826  0.5807 -0.5267  0.5049  0.6200 -0.0489  0.0518
#>  -0.5670  0.5706 -0.4022 -0.0241 -0.3622  0.1000 -0.4205 -0.1231 -0.5702
#>  -0.6678  0.4946 -0.3030 -0.1663 -0.4220  0.2090 -0.3263 -0.5050 -0.3735
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2195  0.4592 -0.0478  0.1195  0.4410 -0.5073  0.0340 -0.4262  0.1882
#>   0.9360 -0.3237 -0.4401 -0.7856  0.3382 -0.6214  0.4504  0.5732 -0.1798
#>  -0.7793  0.3912  0.8176  0.0841  0.3589  0.3480  0.5819  0.0082  0.5706
#> 
#> Columns 10 to 18 -0.2476  0.3014  0.1801 -0.5100 -0.1768  0.1593 -0.6644  0.2487 -0.5500
#>   0.5075  0.0532  0.4135  0.7081  0.0817  0.7164 -0.1203 -0.3392 -0.4722
#>  -0.6615 -0.0461 -0.2895 -0.2795  0.3772  0.3594  0.1727 -0.5953  0.5444
#> 
#> Columns 19 to 20  0.1458  0.5882
#>  -0.2797  0.7658
#>  -0.0837  0.2877
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3351  0.1214  0.1657  0.5643 -0.3227  0.5688  0.2803 -0.3901  0.0689
#>   0.1811  0.0753 -0.1331  0.5435 -0.4387  0.0309  0.0314 -0.6997 -0.2117
#>  -0.5631  0.4729 -0.3712  0.1078 -0.5903  0.4386 -0.4402 -0.4349 -0.4981
#> 
#> Columns 10 to 18  0.0323  0.1766 -0.2813  0.5464  0.1391 -0.3582  0.4815  0.1561 -0.2874
#>  -0.7476 -0.0342 -0.3830 -0.1297 -0.0290 -0.3841 -0.0662  0.3247  0.2535
#>  -0.4305 -0.1874 -0.5456  0.3303  0.0981 -0.5143  0.3971 -0.2588 -0.3115
#> 
#> Columns 19 to 20 -0.0816 -0.1556
#>   0.4229 -0.1177
#>  -0.3151 -0.0236
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward> ]
#>