RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8816 -0.7217  0.8486 -0.9689 -0.1823 -0.8510 -0.0576  0.5167  0.8792
#>  -0.0470 -0.2278  0.6081 -0.7485 -0.5919  0.8644  0.1219 -0.1840 -0.3520
#>  -0.7085  0.4165 -0.2625  0.0965 -0.8807 -0.2298 -0.6753 -0.1026  0.6957
#> 
#> Columns 10 to 18  0.9592 -0.6290 -0.7038 -0.7410  0.4154 -0.0559  0.4427 -0.0228  0.1251
#>   0.8480 -0.4027 -0.6652  0.3508 -0.6244 -0.5909 -0.4759  0.6123 -0.0818
#>   0.4559  0.2106 -0.0897 -0.9728  0.9029  0.5265  0.8723  0.8830 -0.5501
#> 
#> Columns 19 to 20 -0.6709 -0.6607
#>   0.8884 -0.1906
#>   0.3314 -0.4263
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2474 -0.1648 -0.1169  0.1089 -0.4236 -0.1092 -0.8360  0.0185  0.7172
#>  -0.3455  0.3771  0.6441  0.0708  0.0044 -0.0815 -0.1920 -0.1197  0.2384
#>   0.2392  0.8985  0.3966  0.7201  0.0601 -0.0110 -0.5522  0.4203 -0.4313
#> 
#> Columns 10 to 18  0.2762 -0.2229  0.3841 -0.5448  0.6428 -0.2382 -0.1027  0.3626  0.3738
#>  -0.1814  0.3090  0.1756 -0.3390 -0.2967  0.0767  0.2315 -0.2987  0.3455
#>  -0.0327  0.4123 -0.0441 -0.9371  0.6467  0.1707 -0.1626  0.6636  0.0361
#> 
#> Columns 19 to 20  0.3477 -0.7615
#>   0.1001  0.2143
#>   0.3801  0.3507
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1588  0.4284  0.4220  0.3863 -0.4756 -0.3808 -0.2850  0.4246 -0.2526
#>  -0.3940  0.0312  0.2988 -0.3525 -0.1328  0.6316 -0.2190  0.1416  0.2695
#>  -0.3332  0.2568  0.6780 -0.0644 -0.0992  0.2796 -0.2482  0.1275  0.0955
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4288 -0.7309 -0.7008 -0.2499  0.2900 -0.6332 -0.1528 -0.0145  0.0085
#>   0.0863  0.2091  0.2454  0.5721 -0.2460 -0.1656 -0.6436  0.3036 -0.2032
#>  -0.6702  0.5442 -0.4063  0.7169  0.7077 -0.4393 -0.3302  0.7479  0.4653
#> 
#> Columns 10 to 18  0.2920  0.5302  0.2052 -0.2388  0.4458  0.2027  0.0605 -0.2722 -0.4565
#>  -0.4978 -0.5596 -0.8219  0.0181 -0.4102 -0.1144  0.5649  0.5194 -0.0505
#>   0.1159  0.5945 -0.8098 -0.0208 -0.0141 -0.7612 -0.5308  0.6624  0.2056
#> 
#> Columns 19 to 20 -0.1205  0.2969
#>   0.4751 -0.3573
#>  -0.1693  0.0977
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1468  0.1737  0.7240  0.1753 -0.0410 -0.1293 -0.0824  0.1055  0.0047
#>  -0.1694  0.2266  0.5064  0.1669  0.3331 -0.1272 -0.0262 -0.2873  0.1764
#>  -0.0998 -0.1188  0.6788  0.3941  0.1151 -0.4522  0.3963  0.1434 -0.4945
#> 
#> Columns 10 to 18  0.4721  0.1846 -0.1380 -0.4930 -0.3778 -0.2509  0.0852  0.0521 -0.1063
#>   0.2026 -0.0282 -0.1742 -0.5129  0.4840  0.3818  0.5036  0.1818  0.0626
#>   0.1264 -0.0455 -0.0984 -0.5798 -0.3309  0.0680  0.0909  0.5384 -0.0392
#> 
#> Columns 19 to 20  0.5090 -0.2620
#>   0.3546 -0.2184
#>   0.2182  0.0239
#> [ CPUFloatType{2,3,20} ]
#>