RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6155 -0.4206  0.1626 -0.0655  0.6599  0.0900  0.3215  0.0373 -0.1493
#>  -0.4865  0.2712  0.2855  0.3421 -0.3866  0.5953 -0.2262 -0.7027 -0.2764
#>   0.5641 -0.3050  0.8974 -0.1551 -0.5520 -0.6656 -0.6431  0.0098 -0.1932
#> 
#> Columns 10 to 18  0.4331  0.0169  0.1382  0.0030 -0.1747 -0.3437 -0.1904  0.0625 -0.3920
#>   0.2959 -0.9026 -0.4739  0.8718  0.4027  0.7039  0.2569  0.5888  0.9429
#>  -0.0410 -0.1632  0.6873 -0.1469 -0.0659 -0.5976  0.8386 -0.7587 -0.4054
#> 
#> Columns 19 to 20 -0.5878 -0.2937
#>   0.6945 -0.3418
#>  -0.8069  0.1781
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2534  0.2746 -0.0801  0.6777  0.0016 -0.2024  0.1709 -0.2003 -0.0920
#>   0.1945 -0.1711  0.5654  0.4722 -0.5820  0.0575 -0.2037  0.0640 -0.4163
#>  -0.0558  0.1128  0.0201  0.0717 -0.5467  0.2472  0.2652 -0.5250 -0.6988
#> 
#> Columns 10 to 18  0.0683 -0.3008  0.0464 -0.5199 -0.4740  0.0033 -0.1851  0.0389 -0.2609
#>  -0.0711 -0.0924 -0.1659  0.1491 -0.0531 -0.3753  0.5680 -0.5998 -0.3380
#>   0.2806 -0.5883  0.2760 -0.1514 -0.0357  0.0835  0.3468 -0.0189  0.2279
#> 
#> Columns 19 to 20 -0.1947 -0.2226
#>  -0.4875 -0.1208
#>   0.0728  0.2788
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0645 -0.1100 -0.1868  0.5350  0.2627  0.3822  0.1215 -0.4496 -0.2386
#>  -0.5007  0.2698 -0.0977  0.2461  0.5206  0.1311  0.7640 -0.3498  0.2537
#>  -0.3379 -0.1588  0.3765 -0.0528 -0.2281  0.1647  0.5345 -0.0180  0.1721
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0200 -0.7535 -0.0726 -0.8519 -0.3305 -0.1799 -0.3963  0.4661 -0.1660
#>   0.4452  0.2437  0.2133 -0.2776 -0.6923  0.7498  0.2238  0.1784 -0.8012
#>   0.4985  0.6464 -0.0604  0.2464 -0.7261  0.6967 -0.3397 -0.0625 -0.3591
#> 
#> Columns 10 to 18 -0.4728  0.5209  0.1167  0.3590  0.7973  0.4632  0.3259  0.3671 -0.8610
#>  -0.0190  0.0992 -0.2748  0.2645 -0.1444  0.3841 -0.3462  0.3208  0.4493
#>  -0.5000  0.2977 -0.2230  0.1265  0.4141 -0.8269 -0.0647  0.8124 -0.3144
#> 
#> Columns 19 to 20 -0.3810 -0.6592
#>   0.6768  0.3914
#>  -0.5212 -0.4769
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2491  0.0552  0.0748 -0.1681 -0.5075  0.2694  0.2891 -0.6769 -0.3208
#>   0.0229 -0.2046  0.3217  0.1670 -0.4104  0.4131  0.0510 -0.2587 -0.4958
#>  -0.0495 -0.1211 -0.0564  0.3584 -0.2524 -0.1221  0.0209 -0.4822  0.0767
#> 
#> Columns 10 to 18 -0.0850  0.1804  0.3666 -0.3616 -0.4921 -0.3344 -0.0815  0.3553  0.3007
#>   0.3280 -0.4324  0.3227  0.1049 -0.0001  0.0428  0.0865 -0.3711  0.4360
#>   0.3194 -0.6234 -0.1141 -0.5229 -0.2376 -0.0610  0.0488  0.1078  0.3327
#> 
#> Columns 19 to 20  0.5577  0.0407
#>  -0.1329  0.5404
#>  -0.1893  0.2615
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>