RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5680  0.5249 -0.6730  0.7435  0.1527 -0.3243 -0.0063  0.5652 -0.4437
#>  -0.4002  0.7179  0.7725  0.9161 -0.0783 -0.2613  0.8499  0.6717  0.1028
#>  -0.7980  0.9709 -0.1777 -0.2173  0.6226  0.8172 -0.8356  0.2294 -0.8137
#> 
#> Columns 10 to 18 -0.9138 -0.2957  0.5301  0.7489 -0.5610  0.7894 -0.2566 -0.3567  0.5400
#>  -0.1260 -0.8662 -0.0368  0.1034  0.0875 -0.5567 -0.2508 -0.1566  0.3032
#>   0.2150 -0.0253 -0.6627  0.4619  0.3623  0.4933 -0.3645 -0.8513  0.7711
#> 
#> Columns 19 to 20  0.7623 -0.6412
#>   0.7134  0.1100
#>   0.3848  0.8171
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3298  0.0817  0.4602  0.6156  0.1879 -0.0334  0.2458  0.1792 -0.7367
#>   0.3682  0.5843  0.3232  0.7277 -0.1717  0.5129  0.5191  0.3489 -0.5332
#>   0.2768  0.4589 -0.5423 -0.0052  0.1852 -0.1639  0.3290 -0.2236 -0.6100
#> 
#> Columns 10 to 18 -0.0224 -0.0651 -0.3044 -0.5997  0.1329  0.1191  0.1811  0.0350  0.5082
#>  -0.1651 -0.4368 -0.3277 -0.1460  0.2335 -0.1358 -0.1621 -0.5397  0.6030
#>   0.5135  0.3261 -0.4731  0.2353 -0.2939 -0.3430  0.0633 -0.7327  0.0290
#> 
#> Columns 19 to 20  0.6904 -0.1193
#>   0.6821  0.0583
#>   0.7442  0.3909
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4995  0.0705  0.3005  0.1538  0.2768  0.4179  0.7520 -0.0129 -0.2954
#>   0.2066  0.4044  0.4655  0.4965 -0.1081  0.2407  0.4481  0.2029 -0.4686
#>   0.3457 -0.0801  0.3337 -0.2588  0.1645  0.1870  0.2469  0.0414 -0.2602
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2322  0.0470 -0.6079  0.7568  0.5446  0.6077 -0.1534 -0.7110  0.3970
#>  -0.8162  0.2957 -0.8613  0.3044  0.4843  0.6995 -0.1312 -0.1866 -0.6573
#>  -0.0401 -0.4293 -0.2115  0.2247 -0.6078  0.2256 -0.1806 -0.6061  0.2033
#> 
#> Columns 10 to 18 -0.5661 -0.5679 -0.2105 -0.2325 -0.2367 -0.3161 -0.4582 -0.3226 -0.5397
#>   0.3599 -0.4057 -0.1487  0.2873 -0.1091 -0.5418 -0.7171  0.3851 -0.6983
#>   0.4118  0.0897  0.5381  0.2026  0.3443  0.2103 -0.3303  0.1053  0.5313
#> 
#> Columns 19 to 20 -0.7334  0.1216
#>  -0.7632  0.4539
#>  -0.6376  0.2654
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3097 -0.0575  0.2695  0.2809 -0.0635  0.6751  0.2127  0.6777 -0.1115
#>   0.3726 -0.0421  0.5823  0.1482 -0.3683  0.4627  0.6827  0.3693 -0.2338
#>   0.1372 -0.0540  0.4574  0.0913  0.2021 -0.0731  0.4056  0.0812 -0.4875
#> 
#> Columns 10 to 18 -0.5242 -0.3020 -0.4588  0.2654  0.4450  0.1369  0.2545 -0.3692 -0.4562
#>  -0.3293 -0.4228 -0.3661  0.1093  0.3552 -0.1792 -0.0138 -0.3009 -0.4921
#>  -0.3261 -0.1489 -0.4131 -0.2375  0.0396  0.1162  0.3218  0.0200 -0.0664
#> 
#> Columns 19 to 20  0.3907  0.1793
#>   0.3053 -0.2900
#>   0.7282 -0.2189
#> [ CPUFloatType{2,3,20} ]
#>