RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7912 -0.0669 -0.2001 -0.5155 -0.3060  0.3637 -0.2812  0.4709  0.4663
#>  -0.3185  0.6481 -0.0814 -0.9839 -0.1803 -0.0320  0.4598 -0.2816  0.3888
#>  -0.8944 -0.2911  0.6901 -0.2658 -0.6866  0.2863 -0.7115 -0.5962  0.5052
#> 
#> Columns 10 to 18 -0.1990 -0.7818 -0.1776 -0.4658  0.5934  0.0153 -0.0572  0.0389 -0.3098
#>  -0.4865 -0.2121 -0.0014 -0.0350  0.0367  0.4377  0.1852 -0.1112 -0.5427
#>  -0.1074  0.4974  0.2943  0.7032 -0.4675  0.9052  0.8954  0.7927  0.7932
#> 
#> Columns 19 to 20 -0.2896 -0.2791
#>  -0.5187  0.4792
#>   0.3818 -0.7808
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6043 -0.0865  0.1268 -0.5516 -0.5377  0.2300 -0.5577  0.5008 -0.5071
#>  -0.6841 -0.4138  0.1871 -0.2027 -0.6355  0.0877 -0.5830  0.0480 -0.3765
#>  -0.1296 -0.1144  0.7776 -0.2719 -0.3034 -0.2591 -0.0471  0.1048  0.1761
#> 
#> Columns 10 to 18 -0.2071 -0.1117 -0.0928 -0.0346  0.0109 -0.1143  0.0261  0.6157  0.4884
#>   0.4566 -0.1929 -0.3879  0.3877 -0.1848 -0.0075  0.1542  0.0400  0.1186
#>   0.2897  0.1319  0.5067 -0.2405 -0.3908  0.0773 -0.8407  0.0493 -0.2390
#> 
#> Columns 19 to 20  0.0650 -0.2632
#>   0.0424 -0.3324
#>   0.2187 -0.2511
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1135  0.1897  0.5046 -0.6356  0.1465 -0.3898 -0.4462  0.2684  0.4975
#>   0.2540  0.0243  0.7518 -0.1858 -0.2863  0.0771 -0.1367 -0.6099  0.0638
#>  -0.4051 -0.0755  0.7757 -0.2860 -0.0155  0.0160 -0.1545 -0.4456 -0.0691
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1203  0.6931 -0.0820 -0.2641 -0.0883  0.0052 -0.6153  0.6796 -0.1091
#>  -0.2819  0.4437  0.0573 -0.5151 -0.1112 -0.2470  0.2645  0.8577 -0.4002
#>   0.2674  0.4821 -0.2727  0.0831  0.4281  0.0805  0.4370 -0.0587  0.1541
#> 
#> Columns 10 to 18  0.1148  0.5959 -0.1667  0.2876  0.4952  0.3264 -0.5970  0.4190  0.1687
#>   0.4625  0.6784  0.4447 -0.2731  0.1354  0.0214 -0.5545  0.4118  0.1710
#>   0.5767  0.1530 -0.2688  0.2600 -0.1357 -0.6544  0.3435  0.5465 -0.3103
#> 
#> Columns 19 to 20 -0.1981 -0.4300
#>   0.0992 -0.0745
#>   0.5392  0.4941
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3931 -0.4276  0.4665 -0.0283  0.2444 -0.5341 -0.2664 -0.3841  0.1567
#>  -0.6636 -0.0062  0.5826  0.0692 -0.2992 -0.2386  0.0981 -0.3002  0.2056
#>  -0.6206 -0.4986  0.1938 -0.2381 -0.4348 -0.0704 -0.4032 -0.0425 -0.5734
#> 
#> Columns 10 to 18  0.1773  0.2666  0.1276  0.0287 -0.6111  0.0705 -0.4801  0.4877 -0.2593
#>   0.3513 -0.2592 -0.2339  0.0697 -0.7013 -0.2691 -0.6701  0.3281 -0.3750
#>   0.0994  0.2559 -0.1660  0.0935 -0.5473 -0.0210 -0.2467  0.0410  0.2006
#> 
#> Columns 19 to 20 -0.0576  0.0641
#>   0.0953 -0.1482
#>   0.4644 -0.6034
#> [ CPUFloatType{2,3,20} ]
#>