RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2374  0.2698  0.2984  0.1767  0.1652  0.3624  0.1933  0.1783  0.7200
#>   0.5859  0.7596 -0.0669 -0.5489 -0.0815  0.2051 -0.6347 -0.0435 -0.1826
#>   0.8030  0.1827 -0.6063  0.1522 -0.7416 -0.0807 -0.5519  0.2337 -0.4339
#> 
#> Columns 10 to 18  0.7314 -0.0622  0.5076 -0.5384  0.7177  0.2435 -0.5089 -0.5425 -0.8943
#>  -0.1045  0.8305 -0.5669  0.0769 -0.1988  0.5136  0.4418 -0.2163 -0.1254
#>   0.0831  0.0756 -0.0402 -0.0473  0.7657  0.0414  0.6389 -0.0464 -0.5728
#> 
#> Columns 19 to 20 -0.4875 -0.7472
#>  -0.3961 -0.9490
#>   0.7251 -0.2299
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0129 -0.4244 -0.0237 -0.2779  0.0700  0.6183  0.1613 -0.0129 -0.1766
#>   0.4474  0.3595  0.3080  0.1166 -0.1955 -0.1013 -0.0103  0.5344  0.3115
#>  -0.7562 -0.5892 -0.1329 -0.1260 -0.0395  0.3624 -0.1275 -0.2915 -0.0876
#> 
#> Columns 10 to 18 -0.5700 -0.4367 -0.4314  0.1729 -0.2800  0.2671 -0.4982 -0.4198 -0.1099
#>  -0.3954  0.1406  0.1126  0.2246 -0.0964 -0.1591 -0.3418 -0.2029 -0.4791
#>  -0.5345  0.3204 -0.1000 -0.2488 -0.1819  0.2772 -0.4541 -0.2192 -0.3557
#> 
#> Columns 19 to 20 -0.0055  0.2387
#>  -0.2235  0.1441
#>  -0.3745 -0.1711
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1845  0.0095 -0.6050 -0.6343 -0.3690 -0.2225 -0.0657 -0.3657  0.1949
#>  -0.2401 -0.1870  0.3817 -0.1366  0.0219  0.4105  0.3165 -0.2492  0.0759
#>   0.3054  0.5440 -0.3212 -0.2965 -0.3023  0.0956 -0.3383 -0.1755 -0.4868
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5567  0.1418 -0.0162 -0.1383  0.1697  0.8913  0.4845  0.2122  0.3905
#>  -0.2308 -0.8525  0.4718 -0.4208 -0.6898 -0.5626  0.3427 -0.4070 -0.9564
#>  -0.3628 -0.3201 -0.0880  0.3425  0.1775  0.5928  0.6465 -0.1981  0.1632
#> 
#> Columns 10 to 18  0.3786  0.1518  0.4746  0.1933 -0.6118 -0.1102 -0.1350  0.4964 -0.0956
#>  -0.0888 -0.1044  0.5071 -0.3139 -0.6194  0.2940  0.1170  0.2795  0.4260
#>  -0.2127 -0.4998  0.5240 -0.0320 -0.6969 -0.0880  0.3047 -0.1001  0.6364
#> 
#> Columns 19 to 20 -0.2408  0.9423
#>   0.4449 -0.8104
#>  -0.1792  0.7201
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2289  0.3024 -0.1339  0.2690  0.1213 -0.2052 -0.0821 -0.0314 -0.4070
#>  -0.4207  0.1691 -0.0449  0.3583 -0.6799  0.2276  0.5209 -0.4160 -0.2930
#>  -0.4065  0.4399 -0.1048  0.4425 -0.0891 -0.1082  0.2361 -0.1451 -0.3864
#> 
#> Columns 10 to 18 -0.4832 -0.1083 -0.1847 -0.0811  0.0872 -0.2931  0.0009  0.3601 -0.4984
#>  -0.2266  0.0354 -0.5175  0.0054 -0.3186  0.2390 -0.6803 -0.5703  0.0430
#>  -0.6575 -0.4615 -0.1214 -0.2413  0.0003 -0.3220 -0.4606 -0.0319 -0.0762
#> 
#> Columns 19 to 20 -0.2883 -0.3767
#>   0.2940 -0.2177
#>   0.2310 -0.1555
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>