RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8150  0.7550  0.1208  0.0751 -0.7443  0.0803  0.9494  0.7617 -0.7665
#>  -0.8439 -0.1548 -0.4604  0.9028 -0.5244  0.3937 -0.8562  0.8356 -0.3551
#>   0.3992 -0.2703 -0.4959  0.2673 -0.4771  0.4213  0.6907  0.5227  0.0660
#> 
#> Columns 10 to 18 -0.8681 -0.2453 -0.2946 -0.8022 -0.6127 -0.8768  0.1749 -0.8207 -0.5438
#>   0.1634  0.6581  0.9028 -0.0569 -0.9194 -0.5302  0.4127 -0.2180 -0.0497
#>  -0.8220 -0.1418  0.7649  0.0783 -0.3471  0.1839 -0.5090 -0.1038 -0.6123
#> 
#> Columns 19 to 20 -0.0924  0.7287
#>  -0.0410 -0.9046
#>   0.4779  0.5648
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4669  0.0836 -0.0172  0.7517 -0.2062 -0.0736 -0.6057  0.3416  0.0517
#>  -0.5025  0.5657 -0.4265 -0.0997  0.6032  0.7701 -0.6254  0.3972  0.4435
#>   0.0763  0.7113  0.1923  0.3330  0.4359  0.2632 -0.4641  0.6939  0.2480
#> 
#> Columns 10 to 18 -0.0529 -0.0770  0.7000 -0.0063 -0.6247 -0.0004 -0.3800 -0.3423 -0.4267
#>   0.3302 -0.5498 -0.2173 -0.1611 -0.5877 -0.7642 -0.6280 -0.8743 -0.7285
#>   0.1071 -0.3099 -0.3465 -0.0610  0.0106  0.1753 -0.2230 -0.4683  0.2706
#> 
#> Columns 19 to 20 -0.5773  0.1261
#>  -0.3975  0.5506
#>  -0.2481  0.3156
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3896  0.2156 -0.3681  0.2852 -0.0711  0.3401 -0.6405  0.4611  0.5389
#>  -0.3163 -0.0393 -0.4255  0.6206  0.1428  0.3886  0.2815  0.3530  0.3951
#>   0.0337  0.0337 -0.3775  0.4372 -0.0311  0.0978  0.3387  0.1837  0.1034
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5440  0.2755  0.2827 -0.0930 -0.1035  0.7674  0.5918 -0.1872 -0.6461
#>  -0.1881 -0.2348 -0.1136 -0.5256  0.1447  0.3525 -0.0352 -0.0915 -0.8043
#>  -0.2835 -0.7117  0.2326  0.1964  0.9261 -0.1249 -0.0869  0.3735 -0.4286
#> 
#> Columns 10 to 18  0.3118 -0.8817  0.8659  0.0786  0.6515  0.1864 -0.0797  0.0581  0.5879
#>  -0.1374 -0.3113  0.8409  0.0591 -0.0592  0.0639  0.3046 -0.3469  0.0788
#>  -0.5753  0.1472  0.7216 -0.2108  0.6375 -0.0599 -0.4273 -0.2341 -0.2019
#> 
#> Columns 19 to 20  0.6470 -0.1426
#>  -0.2245  0.1946
#>  -0.7453  0.3642
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3246  0.2454 -0.3959  0.5282  0.4372  0.2063 -0.5899  0.5709  0.5750
#>  -0.5177  0.1831 -0.3432  0.4187  0.0553  0.3021 -0.3719  0.4358  0.2572
#>  -0.2281 -0.0733 -0.3482  0.1995  0.6623  0.6770  0.0342  0.3926  0.3029
#> 
#> Columns 10 to 18  0.1546 -0.1408  0.3326 -0.5505 -0.2117 -0.4493  0.0399 -0.7266 -0.2536
#>   0.0098 -0.4235  0.6516 -0.2133 -0.3952 -0.4887  0.1547 -0.1972 -0.4554
#>   0.2980 -0.5381  0.6073 -0.2846 -0.3142  0.2975 -0.3919 -0.1891 -0.1343
#> 
#> Columns 19 to 20 -0.5470 -0.1015
#>  -0.2176  0.0365
#>  -0.6044 -0.0064
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>