RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5244 -0.2535  0.0406  0.2045 -0.2112  0.6093  0.3987  0.7831 -0.5204
#>   0.6688 -0.4064 -0.0919 -0.8451  0.3201 -0.7166 -0.7964 -0.6535  0.8405
#>  -0.7744 -0.0182 -0.9138 -0.1898 -0.0518 -0.6018  0.2274 -0.2686  0.8178
#> 
#> Columns 10 to 18 -0.9247  0.2414 -0.6581 -0.1065  0.1648  0.0935  0.1934  0.6673 -0.2804
#>  -0.6541  0.8377  0.3298  0.6721 -0.9291  0.2169 -0.7129 -0.4905 -0.3200
#>   0.3602  0.3799  0.5280 -0.0755 -0.0426 -0.2805 -0.9638 -0.6681  0.8726
#> 
#> Columns 19 to 20  0.3618 -0.3339
#>  -0.6027 -0.9036
#>   0.7182 -0.6131
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6975  0.3460 -0.5501 -0.0288  0.2251 -0.3924 -0.1098 -0.2563  0.1063
#>  -0.2876 -0.2732  0.2547  0.2463 -0.2974 -0.3917 -0.2391 -0.1131  0.4512
#>  -0.0837 -0.7102 -0.2597  0.0489  0.0666 -0.0248 -0.7426  0.5282  0.5971
#> 
#> Columns 10 to 18 -0.1931 -0.1311 -0.5473  0.0728 -0.0489 -0.7795 -0.6596  0.3463  0.3158
#>  -0.4831  0.2115  0.6006 -0.1740 -0.5741  0.8985 -0.3334  0.1736 -0.0631
#>  -0.3683  0.4106 -0.2899  0.4650  0.1097 -0.0583 -0.4268  0.2667  0.0898
#> 
#> Columns 19 to 20 -0.0291 -0.5179
#>  -0.0183  0.2160
#>   0.0645 -0.6339
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4577  0.3708 -0.2310  0.2527  0.0255 -0.7416  0.1124 -0.2859  0.2189
#>   0.2673 -0.5831  0.0214  0.2129 -0.2485 -0.3354 -0.5153 -0.1253  0.5901
#>   0.2223  0.1913 -0.2963 -0.1751  0.2657 -0.5205 -0.3018 -0.3196  0.4559
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4281  0.1432 -0.2938 -0.5086 -0.0270  0.2010 -0.0851 -0.4021  0.1690
#>  -0.4517  0.6174 -0.4178  0.2294  0.4784  0.4337  0.6789  0.5125 -0.0915
#>  -0.1956  0.3382  0.2054 -0.5828  0.0213  0.2314 -0.3994 -0.2728  0.5148
#> 
#> Columns 10 to 18 -0.4039  0.2123  0.1445  0.1786  0.4074 -0.6056 -0.0944 -0.4716  0.4855
#>   0.6993 -0.3203 -0.1198  0.0552 -0.3464 -0.5576 -0.4399 -0.6028  0.2346
#>   0.2666 -0.2013  0.1647  0.3376 -0.2444 -0.8498  0.0393 -0.5243  0.2897
#> 
#> Columns 19 to 20  0.5188 -0.4162
#>   0.0045  0.6393
#>  -0.5331  0.6383
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3323 -0.3713 -0.1644  0.0876  0.3619 -0.6004  0.2498 -0.1061  0.0655
#>  -0.3132  0.0375 -0.1450 -0.5037  0.0730 -0.2066 -0.3205  0.0470  0.6930
#>  -0.3834 -0.0423 -0.2890 -0.1858  0.4697 -0.3120 -0.2569  0.1116  0.7479
#> 
#> Columns 10 to 18  0.0325  0.2174  0.0440 -0.4632 -0.0225 -0.3277 -0.3290  0.2482 -0.0894
#>  -0.0321 -0.0970 -0.0579  0.1969 -0.4388 -0.0741 -0.2066 -0.5183  0.1702
#>   0.1926  0.4750  0.0465  0.1040 -0.3241 -0.2915 -0.4700 -0.0544 -0.1520
#> 
#> Columns 19 to 20 -0.1577 -0.5453
#>  -0.0833 -0.2185
#>  -0.1009 -0.4705
#> [ CPUFloatType{2,3,20} ]
#>