RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6705 -0.9159  0.7073  0.6994  0.1030 -0.6917 -0.4557 -0.2993  0.4395
#>  -0.3461  0.9050  0.8938 -0.1481  0.8396 -0.5128 -0.2424 -0.3497 -0.1463
#>  -0.1069  0.6702  0.5064 -0.4049  0.7832  0.1331 -0.3571 -0.7416  0.8989
#> 
#> Columns 10 to 18  0.3218 -0.7396  0.4613 -0.8082  0.4324 -0.3966 -0.6342 -0.0299 -0.6675
#>  -0.8062 -0.1869  0.1953  0.9316  0.3864 -0.2437  0.3429  0.7856 -0.1840
#>  -0.8677  0.5200  0.4095  0.5642 -0.7725  0.5194  0.5182  0.0883 -0.2697
#> 
#> Columns 19 to 20 -0.6870  0.4466
#>   0.0184 -0.6164
#>  -0.3075  0.4303
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6459  0.3701  0.1087  0.1259 -0.0288  0.0226  0.1958 -0.0566 -0.1045
#>   0.0791  0.4925  0.4842 -0.2835  0.1052 -0.0740 -0.4816 -0.5687  0.3480
#>   0.2265 -0.1608 -0.6969 -0.2067 -0.4230 -0.4684 -0.3450 -0.6444  0.2010
#> 
#> Columns 10 to 18  0.2631 -0.8247  0.2252  0.3268  0.5837  0.6194 -0.0664  0.3872  0.0379
#>  -0.8073  0.2652  0.7374  0.1311 -0.3098  0.4488 -0.0786 -0.1291 -0.5958
#>  -0.0886 -0.5159  0.5899 -0.5708 -0.6317  0.0891  0.3065  0.2212 -0.2757
#> 
#> Columns 19 to 20  0.1051 -0.6095
#>  -0.1605  0.3306
#>  -0.6163  0.5488
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2613 -0.2754 -0.2317  0.0379 -0.7044 -0.0251 -0.3913 -0.1653 -0.1589
#>  -0.2084  0.1273  0.3834 -0.3219 -0.1827 -0.3915 -0.1292 -0.6521 -0.0337
#>  -0.5864  0.0317  0.2556  0.8357 -0.5972 -0.0884 -0.1024  0.5403  0.2810
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0278  0.1162 -0.1718 -0.4090 -0.6876 -0.3593  0.3149  0.0210  0.1252
#>  -0.3928 -0.1651  0.2441 -0.6552 -0.3740 -0.3014  0.3664 -0.3022 -0.2372
#>   0.5299  0.5076  0.8187 -0.2517  0.1909 -0.9238 -0.7366  0.4240  0.2905
#> 
#> Columns 10 to 18  0.1041  0.2220  0.4214  0.4856 -0.4515  0.4285  0.6520  0.5027 -0.5467
#>   0.5856 -0.0558  0.5115  0.5520 -0.3941  0.1366  0.2959  0.5153  0.2231
#>  -0.3123 -0.2932  0.7342  0.6841 -0.0904 -0.3014  0.5526  0.1474 -0.3114
#> 
#> Columns 19 to 20  0.2972  0.8412
#>   0.1253  0.0798
#>   0.1920  0.5484
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2974  0.2326  0.3329  0.0437  0.1551  0.1113 -0.2644 -0.1629  0.4236
#>  -0.3735  0.5941  0.5746 -0.0639 -0.2860 -0.3169 -0.3210 -0.1660 -0.3701
#>  -0.5283 -0.4898 -0.3443  0.5077  0.0447 -0.1266 -0.5949 -0.3268  0.2365
#> 
#> Columns 10 to 18 -0.1628 -0.3017  0.5381 -0.1451 -0.1798  0.4782 -0.3805  0.0381 -0.0444
#>  -0.2890 -0.5475  0.4704  0.0216  0.0898  0.3152  0.0519  0.3740 -0.6159
#>   0.2326 -0.6236  0.2921 -0.0636 -0.1198  0.2923 -0.0854  0.1069  0.1258
#> 
#> Columns 19 to 20 -0.1090 -0.1941
#>  -0.2324  0.3382
#>   0.2018  0.1265
#> [ CPUFloatType{2,3,20} ]
#>