RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3510  0.1406  0.2664 -0.4188  0.0785  0.3392  0.6429  0.7477 -0.4417
#>  -0.1064  0.7432  0.7969 -0.1512  0.3023  0.3492  0.4203  0.8365 -0.7865
#>   0.6028  0.2845  0.7306 -0.1405  0.0234  0.5650  0.3414 -0.0628 -0.0305
#> 
#> Columns 10 to 18 -0.3124  0.3335 -0.0441 -0.8150  0.7679 -0.4948 -0.8774  0.5768 -0.6781
#>  -0.4486 -0.5450  0.3977 -0.0661 -0.6215 -0.6456 -0.8006  0.5172  0.0487
#>   0.4280 -0.5456 -0.2653  0.4145  0.1633  0.3946  0.1895  0.1798  0.1201
#> 
#> Columns 19 to 20 -0.2578  0.2683
#>   0.0206  0.6022
#>   0.6450 -0.7800
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1330 -0.4491  0.4132  0.0499 -0.2108  0.1074  0.6045  0.4141 -0.4061
#>   0.0695 -0.2953  0.4514 -0.1301 -0.2027  0.0620  0.6068 -0.0902  0.1362
#>  -0.1248  0.6209  0.6094  0.4727 -0.5246  0.3403  0.1578  0.0367  0.3922
#> 
#> Columns 10 to 18 -0.1681  0.3656  0.0162  0.1894  0.3577  0.1649 -0.3798  0.1755  0.5916
#>   0.4722  0.5674 -0.1390  0.1663 -0.0202 -0.0311 -0.1796  0.3219  0.3631
#>  -0.1513 -0.1861  0.0674  0.2585 -0.4351 -0.5633  0.3768 -0.1122  0.3782
#> 
#> Columns 19 to 20 -0.1887 -0.0578
#>   0.1562  0.3058
#>  -0.4408 -0.2361
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1894  0.1211  0.5333  0.0883 -0.2282  0.0336 -0.2604  0.8007 -0.2973
#>  -0.2078  0.2157  0.4963  0.2110 -0.2700 -0.0284  0.3267  0.4080  0.3309
#>   0.3670 -0.2607  0.0538  0.4797 -0.4706  0.3014  0.0657 -0.1026  0.2498
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1677 -0.1326  0.5335  0.3952 -0.4603  0.1295 -0.3591  0.1127  0.7533
#>   0.0051  0.8759  0.3097 -0.1633 -0.4124 -0.6043 -0.1085  0.4601  0.8343
#>  -0.2370 -0.5250  0.3906  0.5919 -0.4541 -0.3606  0.0191  0.3103  0.4299
#> 
#> Columns 10 to 18 -0.1930 -0.5194  0.3185  0.2849 -0.3768 -0.2274  0.0133  0.6539 -0.4573
#>  -0.4053 -0.4410  0.5899  0.6539 -0.5779 -0.2654 -0.1338  0.6087 -0.4970
#>  -0.0797 -0.3909 -0.5259  0.3247 -0.2774 -0.2991 -0.2034  0.2524  0.4507
#> 
#> Columns 19 to 20  0.2456  0.2826
#>  -0.0149 -0.0986
#>   0.0902  0.5218
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1566 -0.0258  0.5518  0.3821 -0.4766  0.4713  0.2366  0.4963 -0.3324
#>   0.4172 -0.0174  0.3355  0.3049 -0.0434  0.3655  0.0876  0.3880 -0.2862
#>   0.2873  0.2679  0.4320 -0.1282 -0.0496  0.2270 -0.1372  0.6552  0.1235
#> 
#> Columns 10 to 18  0.0904 -0.0458  0.1144 -0.3643  0.3129 -0.2667 -0.3000  0.3164 -0.0249
#>  -0.4107 -0.1664  0.0787 -0.1299  0.3933 -0.2880 -0.5601 -0.0933  0.2670
#>  -0.2577 -0.0648  0.2632 -0.0424  0.2040 -0.3362 -0.3926  0.0060  0.2656
#> 
#> Columns 19 to 20 -0.4872  0.0051
#>  -0.5658 -0.3196
#>  -0.6107 -0.1500
#> [ CPUFloatType{2,3,20} ]
#>