RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7254 -0.1486  0.4301 -0.1739 -0.3844  0.1407 -0.5313  0.6306 -0.0781
#>  -0.2301 -0.7521  0.6539  0.5485 -0.8880  0.5622 -0.5910 -0.6038  0.1405
#>  -0.4075 -0.9673  0.8509  0.5217 -0.7441 -0.2505  0.4705  0.8555  0.1919
#> 
#> Columns 10 to 18  0.6744 -0.0227  0.2090  0.8051 -0.3714  0.0470  0.3976 -0.4012  0.3173
#>  -0.4994  0.5103 -0.5962 -0.5063  0.2186  0.2611  0.4001  0.3151  0.3688
#>   0.0341 -0.4891  0.7899  0.4206 -0.4619 -0.2970  0.5085  0.6476 -0.2442
#> 
#> Columns 19 to 20 -0.4606  0.4529
#>  -0.5303 -0.8610
#>   0.7212  0.0819
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0180 -0.6852  0.5252  0.3638 -0.3495 -0.3875 -0.0290 -0.0230  0.0446
#>   0.5941  0.4573  0.4229  0.6616 -0.4018  0.4407 -0.0164 -0.3680  0.3270
#>   0.2750  0.7254  0.4858  0.5238 -0.4951  0.1521 -0.0706 -0.6645  0.5561
#> 
#> Columns 10 to 18 -0.5100 -0.4689 -0.3156  0.2714  0.3613 -0.4899 -0.1181 -0.0885  0.0599
#>  -0.2142  0.3733 -0.3993  0.5257 -0.6765  0.0586  0.2703  0.0098  0.2810
#>   0.2298 -0.3939 -0.6771  0.5932  0.3814  0.0013 -0.1225 -0.1045  0.1480
#> 
#> Columns 19 to 20  0.2491 -0.5255
#>  -0.8286 -0.4014
#>  -0.0822 -0.5618
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1832  0.1497  0.6947  0.1858 -0.4267  0.4677 -0.1506  0.1060  0.0212
#>   0.1816 -0.5830  0.6451  0.3629 -0.4941  0.1457 -0.2104 -0.0416 -0.1177
#>  -0.1294 -0.2207  0.4935  0.2736 -0.0653  0.1253  0.1589 -0.2361 -0.6958
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3954  0.0702  0.0721 -0.4962 -0.5915  0.1298 -0.2035  0.1118  0.0553
#>  -0.5389 -0.8141 -0.1750  0.6200 -0.3248  0.0372 -0.1468  0.1478  0.6235
#>   0.1934 -0.3268  0.2071 -0.6393 -0.4781  0.1891 -0.6619 -0.4354  0.1421
#> 
#> Columns 10 to 18  0.4961 -0.5240 -0.1345 -0.5697  0.1829 -0.0647 -0.2861 -0.0767 -0.0825
#>   0.1067 -0.0683  0.5002 -0.5691  0.3179 -0.5494 -0.2754 -0.3744 -0.0270
#>   0.2415  0.1734  0.7138 -0.4591 -0.1134 -0.1534  0.3126 -0.1529  0.2991
#> 
#> Columns 19 to 20  0.2595 -0.2485
#>   0.0453  0.0733
#>   0.3758  0.0535
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0193  0.0382  0.5919  0.3118 -0.6028  0.5699 -0.0601 -0.5457  0.0638
#>  -0.0562 -0.2377  0.7325  0.6831 -0.3028  0.6505  0.1920 -0.0569 -0.1021
#>  -0.0827 -0.2109  0.6295  0.5422 -0.4738  0.2799  0.1298 -0.4945  0.2803
#> 
#> Columns 10 to 18 -0.0137  0.0345 -0.3222  0.3573  0.1202  0.1155 -0.2225 -0.0052  0.1387
#>  -0.1685  0.3033 -0.1409 -0.2108 -0.2271 -0.1782 -0.1849  0.3955  0.1676
#>  -0.1816 -0.2350 -0.5784  0.2923 -0.3045  0.2452 -0.1377 -0.0701  0.0402
#> 
#> Columns 19 to 20 -0.3664 -0.3157
#>   0.0968 -0.3136
#>   0.0429 -0.5127
#> [ CPUFloatType{2,3,20} ]
#>