RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3534  0.6707 -0.7118 -0.5565 -0.8805  0.8063  0.2726  0.6770 -0.7067
#>   0.8446  0.2597  0.7734  0.6905 -0.7548 -0.5755  0.5824 -0.4899  0.9052
#>  -0.2182 -0.3956 -0.6929 -0.0106 -0.6969 -0.2254  0.4863  0.5462 -0.1727
#> 
#> Columns 10 to 18  0.2115 -0.6659  0.0460  0.8615 -0.8959 -0.6084 -0.3232  0.3442  0.3613
#>  -0.4925  0.3317 -0.1893  0.2225 -0.4761 -0.9582 -0.0727 -0.5376  0.7799
#>   0.5425 -0.4926 -0.8296  0.8338  0.6880  0.1728  0.8557  0.5517 -0.3460
#> 
#> Columns 19 to 20 -0.3899 -0.7219
#>  -0.5048  0.2130
#>   0.3205 -0.5958
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2088 -0.5963 -0.4526  0.3858  0.3027 -0.3827 -0.7535 -0.0244 -0.7237
#>  -0.7558 -0.1327 -0.3247 -0.4015  0.4100  0.3978 -0.3780 -0.7321 -0.5425
#>  -0.0741  0.3142 -0.5080  0.6637 -0.6325 -0.5586 -0.2085 -0.4298 -0.3912
#> 
#> Columns 10 to 18 -0.8204  0.0809 -0.3995  0.2378 -0.0786 -0.1879  0.6240  0.1874  0.4337
#>  -0.5122  0.1286  0.3127  0.0787  0.2217  0.0252 -0.0843 -0.3028 -0.3301
#>   0.0098 -0.4248 -0.2912  0.7946  0.1635 -0.0671  0.5946 -0.1990  0.1223
#> 
#> Columns 19 to 20  0.4453  0.3849
#>   0.1287 -0.0946
#>   0.7725  0.0376
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2461  0.4872 -0.6413  0.1185 -0.3472 -0.4881 -0.3736  0.2424 -0.0051
#>   0.3979 -0.5526 -0.2894 -0.1351  0.2413 -0.1326 -0.0636  0.1138 -0.2105
#>   0.4818  0.1803 -0.2190  0.0722 -0.0428 -0.2757 -0.6224  0.2147  0.4548
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6665  0.1620  0.3280  0.5544 -0.6634  0.1897  0.0852 -0.8376 -0.0156
#>   0.1788  0.7001  0.4703  0.4610 -0.0669 -0.3911 -0.4067 -0.0018  0.2069
#>   0.4712  0.3196 -0.2638  0.3917 -0.2740  0.0085  0.3287  0.2739  0.5316
#> 
#> Columns 10 to 18 -0.0613  0.8206 -0.4700 -0.7360  0.3315 -0.6520 -0.7586 -0.2092  0.4995
#>   0.6377  0.0217  0.3208 -0.3566  0.6542  0.4999 -0.3146 -0.1101 -0.6026
#>   0.1008  0.6182  0.2303 -0.1034  0.3199  0.2026 -0.2174 -0.3559  0.3411
#> 
#> Columns 19 to 20 -0.1861  0.4624
#>   0.4878 -0.0320
#>  -0.0602 -0.0836
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5783  0.1492 -0.3132 -0.1929  0.5368  0.4966  0.0080 -0.2489 -0.4109
#>  -0.1707 -0.2069 -0.2903  0.3500 -0.3766 -0.2342  0.0704 -0.3402  0.1398
#>  -0.1872 -0.1920 -0.4963  0.2239  0.2105  0.0278 -0.2604  0.0271 -0.1240
#> 
#> Columns 10 to 18 -0.5454  0.0835  0.1285  0.3502  0.2824 -0.1780  0.1754 -0.1713 -0.5722
#>  -0.3260  0.5015  0.4313  0.4391  0.6181 -0.6505  0.3724  0.0494  0.2159
#>  -0.3827 -0.1798  0.1122  0.1000  0.5037 -0.5364  0.2057 -0.1668 -0.1587
#> 
#> Columns 19 to 20  0.2474 -0.0972
#>   0.0338 -0.1455
#>   0.1888 -0.3866
#> [ CPUFloatType{2,3,20} ]
#>