RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4851  0.4566 -0.1978 -0.0187 -0.6524 -0.3286 -0.6617  0.2513  0.6108
#>  -0.5832  0.6981  0.3989 -0.2609 -0.7542  0.7748 -0.1162 -0.5019  0.7400
#>  -0.5995  0.5988  0.6559  0.1581  0.1759 -0.1027 -0.3062 -0.0784 -0.7583
#> 
#> Columns 10 to 18 -0.2058  0.3806 -0.4139  0.3469 -0.1143 -0.7264  0.2980  0.7077  0.6498
#>  -0.5905  0.7212  0.9302  0.0301  0.8241 -0.7702 -0.0654  0.0185  0.7899
#>  -0.2956 -0.3377 -0.1309  0.0835  0.3074 -0.1840  0.0297 -0.3653  0.6004
#> 
#> Columns 19 to 20  0.0579  0.7175
#>   0.7927 -0.4404
#>  -0.3624  0.6011
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1194 -0.3865  0.3868  0.5613 -0.6499 -0.2211 -0.0916 -0.1116  0.3387
#>  -0.6477 -0.0087  0.8184  0.6072  0.5051 -0.7170  0.1204  0.2301 -0.2534
#>  -0.7319  0.0543  0.0604 -0.0549 -0.2623 -0.2760  0.1935  0.4694  0.2046
#> 
#> Columns 10 to 18 -0.4594  0.7153 -0.0705  0.1019  0.2005  0.0063  0.0915  0.2895 -0.0890
#>  -0.3348  0.1662 -0.1716  0.4693  0.0236  0.2213 -0.4799  0.2580 -0.5218
#>  -0.0214  0.6116  0.1603 -0.4806 -0.4223 -0.4036  0.3587  0.2769  0.3130
#> 
#> Columns 19 to 20  0.0494  0.6095
#>  -0.2490  0.4706
#>   0.4601  0.5311
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0922 -0.0788  0.0176  0.1740 -0.4930  0.0279  0.1310 -0.0994 -0.0954
#>   0.5056 -0.5107 -0.4862 -0.0784 -0.3062 -0.5126  0.1425  0.1652  0.2720
#>  -0.1888 -0.2783  0.0463  0.2183 -0.2390 -0.4818  0.0502 -0.1512 -0.1275
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1149 -0.4932  0.5818  0.0658  0.0456 -0.2214  0.5425  0.3399  0.5378
#>  -0.5546  0.2879 -0.3855  0.1804  0.5209  0.3792  0.4351  0.5143 -0.0844
#>  -0.1539  0.3044  0.1899 -0.6274  0.7672  0.0026  0.2509  0.6173  0.0620
#> 
#> Columns 10 to 18 -0.5311  0.6601 -0.3641  0.6497  0.1468 -0.2963  0.6311  0.6652  0.4798
#>   0.1134 -0.7621  0.6008 -0.0094  0.6425 -0.1578  0.6115 -0.4597 -0.0567
#>  -0.3232 -0.3109  0.4153  0.2074 -0.0084  0.7599  0.5277 -0.0909  0.7300
#> 
#> Columns 19 to 20 -0.3300  0.4869
#>  -0.5131 -0.5962
#>  -0.0150  0.1299
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3156  0.2488  0.4676  0.2309 -0.6373 -0.6629  0.1659  0.0050  0.0437
#>  -0.4583 -0.4587 -0.0072  0.4022 -0.1923 -0.4092 -0.2731  0.3554 -0.0857
#>  -0.0966 -0.5697  0.2670 -0.0280 -0.1064 -0.6008 -0.5050 -0.2751  0.6118
#> 
#> Columns 10 to 18 -0.4797 -0.2386  0.1673  0.3083  0.4212 -0.2458 -0.6412 -0.0560 -0.2361
#>  -0.1203  0.2139  0.2434  0.5153 -0.0492 -0.5303  0.3825  0.0239 -0.0500
#>  -0.4620 -0.0744  0.3640  0.4721 -0.0267 -0.3564  0.4496  0.4937 -0.0127
#> 
#> Columns 19 to 20  0.0679  0.5145
#>  -0.0611  0.5631
#>   0.1114  0.4361
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>