RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0538 -0.0973  0.1179  0.2235 -0.6345  0.1884 -0.7703  0.6531 -0.4569
#>  -0.3553  0.4359  0.6743 -0.6284  0.8402 -0.2570  0.5619 -0.1295 -0.7527
#>   0.5302 -0.1292  0.8129 -0.0240  0.8884 -0.6946  0.0973  0.1374 -0.8274
#> 
#> Columns 10 to 18  0.6086  0.8402 -0.7238  0.6095  0.5974 -0.7890  0.6929  0.7740 -0.6833
#>   0.3839 -0.0522  0.4505 -0.5147 -0.7339  0.4365 -0.7592 -0.4012  0.4697
#>   0.9189  0.0739  0.1264  0.2810 -0.4351 -0.3261  0.6614  0.8269 -0.7598
#> 
#> Columns 19 to 20 -0.3743 -0.8566
#>   0.1333  0.7654
#>   0.0046 -0.7953
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5044  0.0159 -0.3270  0.3153  0.0801 -0.3614 -0.0119 -0.3398 -0.1270
#>  -0.1402 -0.2623  0.4123  0.3718  0.6430 -0.0196  0.2898  0.4808 -0.3396
#>   0.8560  0.1080 -0.1698  0.4308  0.4841 -0.1742  0.0957 -0.5200  0.0434
#> 
#> Columns 10 to 18 -0.1394  0.5072  0.3402  0.1451 -0.5463 -0.2610  0.0819  0.0890  0.0641
#>   0.4347 -0.6763  0.0712 -0.0348 -0.1082  0.3437 -0.1902  0.5821 -0.0913
#>  -0.1861 -0.1686  0.5640  0.0522  0.1800  0.2900 -0.0631  0.5462 -0.2396
#> 
#> Columns 19 to 20 -0.3088 -0.1725
#>  -0.0094  0.3285
#>  -0.6802 -0.4496
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4796  0.5057 -0.4438  0.4427  0.5683 -0.2097 -0.0368  0.0304 -0.4495
#>   0.3290  0.0044  0.0461 -0.0954  0.5783  0.0055  0.3041  0.0159 -0.0351
#>   0.2653 -0.0191  0.3369  0.4061  0.0025  0.4249 -0.0340 -0.3707  0.3771
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1158 -0.3346  0.4847  0.1262 -0.1386  0.4390 -0.7204 -0.1033  0.0396
#>   0.0051  0.5298 -0.0533 -0.0865 -0.3318  0.0355 -0.2162 -0.1555 -0.2975
#>  -0.4016  0.4394  0.1572 -0.0115 -0.5421  0.1210 -0.3048 -0.5680 -0.2925
#> 
#> Columns 10 to 18  0.3610  0.4089 -0.1992 -0.0842  0.5522  0.0342  0.2992  0.2605 -0.6060
#>   0.0202  0.3978  0.1130  0.2994  0.3042  0.6693  0.4968 -0.5255  0.1442
#>  -0.1355 -0.1351  0.4333 -0.1341  0.1757  0.1192 -0.4351 -0.0477 -0.3798
#> 
#> Columns 19 to 20  0.3157  0.1296
#>  -0.3417  0.2784
#>   0.2464 -0.3608
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2233  0.4118 -0.1665  0.1726  0.4816  0.1820  0.2478  0.1653 -0.2332
#>   0.4350  0.6002  0.3245  0.2062  0.3522 -0.1882  0.2354 -0.1405 -0.1212
#>   0.0260  0.0965  0.0249  0.4836  0.6946  0.3784  0.5363  0.2982 -0.4086
#> 
#> Columns 10 to 18 -0.0202  0.1860 -0.1452 -0.3577 -0.3992  0.1985 -0.0594 -0.0007 -0.4009
#>  -0.0948 -0.2070  0.5542  0.0369  0.1089  0.2014  0.0214  0.5831 -0.4373
#>  -0.2933 -0.0257  0.1508 -0.1861  0.0067  0.4992 -0.2331  0.2366 -0.1011
#> 
#> Columns 19 to 20 -0.1207 -0.2638
#>  -0.1634 -0.0429
#>  -0.2872  0.0552
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>