RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4283  0.3982 -0.8821  0.8006 -0.0950 -0.6206  0.8121 -0.1823 -0.1259
#>  -0.0219  0.4353 -0.3542  0.6528 -0.8792  0.0266 -0.0877  0.0255  0.2844
#>  -0.2626 -0.0375 -0.7282 -0.0301  0.4584  0.4145 -0.1591 -0.4772 -0.6051
#> 
#> Columns 10 to 18  0.5288  0.5544  0.5070  0.9095 -0.4240 -0.2927  0.4969 -0.8014  0.1620
#>   0.6658  0.2574  0.0073  0.1712  0.2255 -0.5394 -0.1964  0.3365 -0.4049
#>  -0.0789  0.0004 -0.7306  0.6447  0.5423  0.4499  0.2549  0.3500 -0.2587
#> 
#> Columns 19 to 20  0.8932 -0.5054
#>   0.5524  0.3251
#>  -0.6128  0.2860
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5530  0.2659 -0.0319  0.3027 -0.1108 -0.2096 -0.4098  0.0847  0.0465
#>  -0.7119  0.2369  0.0770  0.4144 -0.0724 -0.6531 -0.4692  0.1461 -0.3243
#>   0.1574  0.0140 -0.0512  0.0163 -0.0894 -0.1924 -0.7670  0.2480 -0.2098
#> 
#> Columns 10 to 18  0.0597  0.2342 -0.1273 -0.0742 -0.3765 -0.4895 -0.0037 -0.6667  0.2930
#>   0.1208  0.5694  0.1046  0.0913 -0.6474 -0.4738  0.4663  0.4632  0.6306
#>   0.2099  0.1122 -0.0385  0.3803 -0.7538 -0.3118 -0.0166 -0.2307 -0.3209
#> 
#> Columns 19 to 20  0.2120  0.3955
#>  -0.0675  0.6386
#>  -0.4666 -0.0459
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5474 -0.1563 -0.1843  0.6374 -0.4239 -0.2544 -0.5092  0.0868  0.3458
#>  -0.2427  0.2413  0.0652  0.3130 -0.2645 -0.1434 -0.6983  0.2513  0.3599
#>  -0.2010  0.3666 -0.5518  0.6441 -0.0885 -0.5003 -0.4792 -0.1025  0.1236
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3887 -0.1932 -0.6803 -0.4546 -0.8865  0.4889 -0.3860 -0.4204  0.7943
#>   0.1780  0.2100  0.4882 -0.7161 -0.1261 -0.4833 -0.6669  0.2745  0.0354
#>  -0.0458 -0.3807  0.3721 -0.2892 -0.0328 -0.6164 -0.4046  0.4610  0.3196
#> 
#> Columns 10 to 18 -0.3691 -0.3430  0.1329  0.7004  0.3164 -0.3498  0.0460  0.0695 -0.5673
#>  -0.3539 -0.5001  0.6486 -0.6556 -0.4382  0.1581  0.1733  0.6175  0.0999
#>  -0.7207  0.1868  0.4523  0.3284 -0.6473 -0.0810 -0.0181  0.4399  0.5038
#> 
#> Columns 19 to 20 -0.0649  0.2269
#>  -0.1912 -0.5930
#>  -0.0832 -0.2788
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3206  0.1045 -0.4433  0.0562 -0.2159 -0.1083 -0.3244  0.0824  0.4484
#>  -0.0522 -0.2018 -0.2499  0.7614 -0.2309 -0.5304 -0.3286  0.1045  0.1620
#>  -0.2423  0.0435 -0.4965  0.5738  0.0520 -0.4878 -0.7667  0.0941  0.2792
#> 
#> Columns 10 to 18 -0.5735  0.3879 -0.0009 -0.2266 -0.6980  0.1597  0.2405 -0.3074 -0.0162
#>  -0.3382  0.3187 -0.3238  0.1547 -0.7563  0.6364  0.7464  0.1344 -0.4263
#>  -0.3547  0.4546 -0.3471  0.3373 -0.8319  0.3002  0.8306  0.5378 -0.4424
#> 
#> Columns 19 to 20  0.2656  0.2075
#>   0.0293  0.5886
#>  -0.0119  0.5906
#> [ CPUFloatType{2,3,20} ]
#>