RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8113 -0.3902  0.5428  0.0866 -0.7280  0.5012  0.7773  0.7873  0.2511
#>   0.8764  0.8719  0.5368 -0.0070 -0.3749  0.2764  0.4224  0.9145  0.1157
#>   0.4398  0.7695  0.5163  0.3164  0.7996 -0.8878  0.4127 -0.2148  0.2825
#> 
#> Columns 10 to 18 -0.6086 -0.4666 -0.7683  0.6305  0.0976 -0.5947 -0.5667  0.7840  0.5643
#>   0.3102 -0.2726  0.7961  0.0057  0.6387  0.7130 -0.5764 -0.6880  0.2962
#>  -0.1451 -0.1170 -0.0911 -0.2328  0.2113 -0.1827 -0.0213 -0.4862 -0.6424
#> 
#> Columns 19 to 20 -0.7693 -0.0704
#>  -0.7323  0.1630
#>   0.5873 -0.4075
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.6708 -0.0796  0.1280  0.0840  0.3204 -0.3965 -0.2777  0.7091  0.2086
#>  -0.2320 -0.0096  0.5532  0.6193  0.1890  0.3191 -0.1108  0.3563  0.3988
#>  -0.3385  0.2635  0.2666 -0.2518  0.4697  0.3023  0.0457 -0.0961 -0.2977
#> 
#> Columns 10 to 18  0.8570 -0.4870  0.5856  0.4964 -0.1931  0.0601  0.1088  0.2062  0.1247
#>   0.7629 -0.6210  0.5422  0.5157  0.4208 -0.0155 -0.7286  0.7132  0.3491
#>   0.8131  0.3452  0.5512  0.3443  0.2588 -0.1895 -0.4480  0.4422 -0.3651
#> 
#> Columns 19 to 20  0.1377  0.5009
#>   0.2323  0.8334
#>   0.0830 -0.1838
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2811  0.2708  0.3082 -0.4576  0.5246  0.5825  0.6245 -0.2201 -0.5070
#>  -0.2717 -0.1562  0.4500 -0.1172  0.7432 -0.3327 -0.1070  0.1305  0.7264
#>  -0.1261 -0.1445 -0.0098 -0.0454  0.8072  0.3331  0.1305  0.5434  0.1310
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.6458  0.5926 -0.0782  0.8025 -0.3357 -0.2773 -0.3679  0.7805  0.4852
#>   0.0122 -0.6303  0.1286  0.6224 -0.6643  0.3296 -0.6075  0.7987 -0.1927
#>  -0.6106 -0.2475  0.1702  0.4996 -0.3849 -0.2442 -0.7127  0.6402  0.2419
#> 
#> Columns 10 to 18  0.6154  0.6041  0.4736 -0.1552 -0.8789 -0.1055  0.5807 -0.8074 -0.0015
#>  -0.4471  0.0669 -0.4068  0.6506 -0.1401  0.0614  0.0352  0.6857 -0.6204
#>  -0.2871  0.1282 -0.4548  0.3658 -0.6094 -0.5031  0.4761 -0.0633 -0.3203
#> 
#> Columns 19 to 20 -0.6381  0.8331
#>   0.0151  0.7609
#>  -0.2688  0.5990
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1191 -0.1933  0.0017  0.1215  0.7390 -0.3349  0.1044  0.5677  0.1473
#>  -0.1685  0.3711  0.1737 -0.0695  0.5215  0.4216  0.2466 -0.0833 -0.4097
#>  -0.3070  0.0156  0.1892 -0.4817  0.2240  0.3239  0.4154  0.2826 -0.5361
#> 
#> Columns 10 to 18  0.3768 -0.0552  0.0466  0.3980  0.0360 -0.3008 -0.5043  0.1576  0.7596
#>   0.2394 -0.2160  0.4691  0.3921  0.6572 -0.5203 -0.0775  0.1559  0.6070
#>   0.2018 -0.3363  0.2037  0.4455  0.6252 -0.4405 -0.0757 -0.0095  0.5231
#> 
#> Columns 19 to 20  0.1921  0.0680
#>   0.3526 -0.1043
#>   0.2752  0.0302
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>