RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4836 -0.3131  0.5964  0.0818  0.0315 -0.1279 -0.3793 -0.6406 -0.3423
#>   0.2743 -0.5668  0.1477 -0.7491  0.4119  0.0993 -0.4686 -0.6278 -0.8006
#>  -0.3211 -0.5410 -0.0783  0.6918  0.1946  0.7411  0.4900  0.6982  0.3187
#> 
#> Columns 10 to 18 -0.6637 -0.4242 -0.1175  0.1569  0.3750 -0.3971 -0.5770  0.5343  0.3319
#>   0.2942 -0.8818  0.4784 -0.1558 -0.4007 -0.3713  0.7276 -0.7129  0.3596
#>   0.8762  0.5072 -0.5648 -0.2582 -0.4731  0.4414 -0.4322  0.7499 -0.7844
#> 
#> Columns 19 to 20  0.2622 -0.0263
#>   0.4053 -0.5585
#>  -0.9726  0.1045
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3764 -0.1194 -0.3395 -0.4376  0.6716  0.6106 -0.1151 -0.2829 -0.1817
#>  -0.1343 -0.3050  0.1119  0.4513 -0.2148  0.1723  0.0350 -0.1059  0.2392
#>   0.8586 -0.2772 -0.2191 -0.7404  0.1806  0.4067 -0.0940  0.0875  0.5230
#> 
#> Columns 10 to 18  0.2869  0.1453  0.4439  0.1881  0.0155 -0.1316 -0.5042 -0.1206  0.3867
#>  -0.2677  0.1672  0.1737 -0.1717  0.0263 -0.0129 -0.5624  0.0638  0.3662
#>   0.4657 -0.1136 -0.6789  0.0262  0.0893 -0.1380  0.3015 -0.0763 -0.1737
#> 
#> Columns 19 to 20 -0.2489 -0.1078
#>  -0.5290 -0.1350
#>  -0.0885  0.2962
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2445  0.0403  0.1312 -0.2490 -0.0680  0.2989 -0.2402 -0.2328 -0.1712
#>   0.4907 -0.1354  0.1436 -0.3575  0.2373  0.3404  0.0219 -0.3142 -0.1676
#>  -0.1630  0.0545  0.1243 -0.0400  0.4018  0.5518 -0.2206 -0.3580  0.1190
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5788 -0.3065 -0.4703 -0.4920  0.4748  0.2734  0.5138 -0.2322 -0.5470
#>  -0.0589 -0.6952 -0.4978  0.4626 -0.6192 -0.7347 -0.1081  0.3665  0.4532
#>  -0.4458 -0.0838 -0.4440  0.3822 -0.7495  0.2281 -0.5386  0.0122  0.4788
#> 
#> Columns 10 to 18 -0.4528  0.5320 -0.1284  0.8039 -0.0983  0.5273  0.1323  0.0621  0.1025
#>  -0.2366 -0.7381 -0.2389  0.4406 -0.4165  0.1818  0.2524 -0.4420  0.6779
#>   0.1357 -0.0277 -0.9021  0.4190 -0.1486  0.8433 -0.6001 -0.4986  0.7489
#> 
#> Columns 19 to 20  0.3750  0.7312
#>  -0.2878 -0.4144
#>  -0.6069 -0.4231
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4234 -0.0980 -0.2824 -0.5175  0.0189  0.5173  0.0157 -0.4128  0.0396
#>   0.0452 -0.3914 -0.2140 -0.3627  0.4160  0.6760 -0.1214 -0.2065  0.1482
#>   0.0694 -0.0727  0.1904 -0.5366  0.3192  0.1214 -0.3401 -0.5496  0.2087
#> 
#> Columns 10 to 18  0.0675 -0.0072  0.2139 -0.1022  0.3385  0.4549 -0.2114 -0.3042  0.1405
#>   0.5046 -0.2705  0.0639 -0.1278 -0.0264 -0.2233 -0.2376 -0.1524  0.3002
#>   0.1088 -0.0866 -0.5023 -0.0854 -0.0979 -0.2855  0.1568  0.0157  0.4881
#> 
#> Columns 19 to 20 -0.2851 -0.3472
#>  -0.1004  0.1534
#>   0.1420  0.3106
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>