RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5181  0.1660 -0.2946 -0.5048  0.7662  0.1119  0.1582  0.8809  0.1633
#>  -0.1874 -0.9416  0.7877  0.3066  0.7483  0.5770  0.6506 -0.7842 -0.2861
#>   0.5698  0.9329  0.5479 -0.7861 -0.2044 -0.7215  0.0633  0.4481 -0.2764
#> 
#> Columns 10 to 18 -0.9178 -0.7399 -0.4208 -0.7286 -0.0782  0.2032  0.9409 -0.6569 -0.8524
#>   0.7482  0.2525 -0.0271 -0.7321 -0.3327 -0.2169 -0.7317 -0.7985 -0.3452
#>   0.3585  0.3167 -0.4219  0.1327 -0.0770 -0.1550 -0.2624  0.4779 -0.2848
#> 
#> Columns 19 to 20 -0.5795  0.7591
#>  -0.5014  0.0342
#>   0.5897 -0.2574
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2712 -0.2378 -0.6203 -0.1157  0.3348  0.0331  0.5054 -0.1757 -0.5581
#>   0.4989 -0.0064  0.4275 -0.0896 -0.3543 -0.1439  0.4198  0.6129 -0.3926
#>   0.5189  0.3190  0.5827  0.5636 -0.4144  0.1067  0.3785  0.5868 -0.1456
#> 
#> Columns 10 to 18  0.1898 -0.1294 -0.4414 -0.1167 -0.5531 -0.1208  0.3830 -0.7047 -0.5072
#>  -0.5373 -0.5280 -0.5264 -0.4247  0.5073 -0.5928 -0.1112 -0.0286 -0.2005
#>  -0.1378  0.2315  0.3866  0.1375  0.3517  0.3440  0.2838 -0.1594  0.1995
#> 
#> Columns 19 to 20  0.1243 -0.5631
#>   0.0741  0.6312
#>   0.6432 -0.2776
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.6679  0.2812 -0.1514 -0.0022 -0.1968 -0.0970  0.2737  0.5754 -0.2781
#>   0.5942  0.0147  0.1714  0.5999  0.0687  0.6481  0.6783  0.2780 -0.2416
#>  -0.0927  0.0550  0.5988  0.4330  0.3202  0.5459 -0.1983  0.2751 -0.3014
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0668 -0.3016 -0.4529  0.1479 -0.7849  0.7236 -0.7519 -0.2794 -0.0147
#>  -0.1648 -0.2519 -0.6530  0.7002  0.1120  0.7783 -0.8105 -0.7343  0.3654
#>  -0.0838 -0.3649 -0.5190  0.8637  0.9016 -0.1991 -0.0584  0.2436  0.1275
#> 
#> Columns 10 to 18  0.4584 -0.0144  0.3436  0.2647  0.0936  0.2454 -0.3015  0.1008  0.5574
#>  -0.5608 -0.6755  0.0308  0.1280  0.2647 -0.4396 -0.5011 -0.4011  0.1365
#>  -0.5867 -0.2194  0.2829  0.6829 -0.4574 -0.5200 -0.6546 -0.1520 -0.3511
#> 
#> Columns 19 to 20 -0.1284  0.3232
#>  -0.7153 -0.1283
#>   0.0101 -0.2500
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0846 -0.0233  0.4553  0.1048 -0.0149 -0.0019  0.1312  0.2347 -0.3031
#>   0.2437  0.1602  0.3723  0.1152 -0.1247  0.5164  0.2914  0.7575 -0.2501
#>   0.6388  0.4776  0.2058  0.1673 -0.2585  0.4811  0.1364  0.7452  0.0073
#> 
#> Columns 10 to 18 -0.6329 -0.3808 -0.2268  0.0123  0.0289 -0.1568  0.2944 -0.0657 -0.1215
#>  -0.4761 -0.3803 -0.4688 -0.3790  0.2585 -0.0328  0.3418 -0.0878 -0.5654
#>  -0.4263 -0.1850 -0.2459 -0.1199  0.2196 -0.0631  0.4563 -0.4298 -0.4694
#> 
#> Columns 19 to 20  0.6415 -0.2207
#>   0.6229  0.0268
#>   0.6822  0.1338
#> [ CPUFloatType{2,3,20} ]
#>