RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3440  0.5490 -0.5129 -0.4637  0.6645 -0.5124 -0.0644 -0.3901 -0.4145
#>  -0.2264  0.8656 -0.7932 -0.5517 -0.6375 -0.5986 -0.9024  0.3601  0.6141
#>   0.0104  0.6740 -0.5276  0.5604 -0.4171  0.3743 -0.5419  0.0300  0.8419
#> 
#> Columns 10 to 18 -0.5227 -0.0499  0.8347  0.0630  0.5179  0.0906 -0.2113 -0.4755 -0.5402
#>   0.2309 -0.3686  0.4129  0.4040  0.4465 -0.5260 -0.6178 -0.4724  0.3249
#>   0.4092 -0.0919 -0.5215 -0.2046  0.8453 -0.1082 -0.1494  0.6683 -0.5016
#> 
#> Columns 19 to 20  0.4762 -0.1652
#>  -0.8083 -0.2272
#>  -0.8782  0.6903
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2522  0.3066 -0.1267  0.3131 -0.1183 -0.2721  0.4709  0.3429 -0.1427
#>   0.5616 -0.2280  0.5041  0.3656 -0.4751  0.8034 -0.5590  0.0749 -0.1751
#>  -0.2765  0.4121 -0.1534  0.1425  0.4043 -0.2909  0.4674 -0.2214 -0.3679
#> 
#> Columns 10 to 18  0.2691 -0.1869 -0.1886 -0.3001 -0.3026 -0.0357 -0.4509  0.5231  0.5588
#>  -0.0116 -0.2242 -0.2082 -0.4061 -0.1247  0.4264 -0.5992 -0.0943  0.6697
#>  -0.3457  0.0923  0.3239  0.3910 -0.2377 -0.2041  0.5565  0.3697  0.2888
#> 
#> Columns 19 to 20  0.3177  0.2856
#>   0.5965 -0.4909
#>   0.2722  0.0939
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0054 -0.2552  0.4303  0.1951 -0.4443 -0.2836 -0.0142  0.2933 -0.0068
#>   0.3282 -0.5673 -0.2642 -0.3724 -0.7870  0.1172  0.3255 -0.4424  0.1737
#>  -0.2772 -0.0825  0.1273  0.2843 -0.0350 -0.3643 -0.0690  0.4711 -0.1044
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8388 -0.5315 -0.6012 -0.9112  0.3843  0.0948  0.7035  0.1950 -0.4018
#>   0.3176  0.2488  0.6763 -0.7127 -0.2935  0.6509  0.2374  0.0093  0.2432
#>   0.1258 -0.0554  0.4217 -0.7103  0.0001  0.4249  0.6238  0.6056 -0.1027
#> 
#> Columns 10 to 18  0.7286 -0.5764 -0.5220 -0.0777  0.2245 -0.3814 -0.4524  0.0127  0.7791
#>  -0.2452 -0.3444  0.2222 -0.1517  0.6378  0.1352  0.4335  0.5925 -0.2569
#>  -0.8431 -0.1373 -0.0341  0.3043 -0.3223  0.8322  0.3455  0.0628  0.7324
#> 
#> Columns 19 to 20  0.0010 -0.7176
#>   0.2583 -0.0235
#>   0.1049  0.3135
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2334 -0.1790  0.4898 -0.1187 -0.7085  0.5911  0.1202 -0.5977 -0.0638
#>   0.2019 -0.5407 -0.2362 -0.3527 -0.4409  0.0269  0.4510 -0.6061 -0.2510
#>   0.4494  0.0086  0.1758 -0.2384 -0.5089  0.1030  0.4545 -0.1447 -0.1469
#> 
#> Columns 10 to 18 -0.5602  0.2983 -0.3294 -0.2027  0.2444 -0.3150  0.1091 -0.1375  0.1621
#>  -0.0762  0.3347 -0.3800 -0.1631  0.2949 -0.3563  0.3169  0.2273  0.4566
#>  -0.3938  0.3918 -0.5518  0.4453 -0.0709 -0.2781  0.2704  0.3406  0.6321
#> 
#> Columns 19 to 20  0.7426 -0.3034
#>   0.3816  0.0264
#>   0.5087  0.2277
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>