RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1687 -0.1026  0.2076 -0.0463 -0.0673  0.1754 -0.3603  0.3728  0.5044
#>   0.5666  0.2571  0.7380  0.2627  0.0505  0.9429  0.0037 -0.9264  0.4704
#>  -0.4006  0.0842  0.4504  0.2726  0.7777  0.4103  0.5824 -0.6212  0.3780
#> 
#> Columns 10 to 18 -0.2904  0.1692 -0.1981  0.1248 -0.6368 -0.7868  0.4133 -0.5504 -0.1105
#>  -0.2652 -0.5537 -0.0024  0.5588 -0.1431  0.7368  0.2550  0.3890  0.1350
#>  -0.2419 -0.2009  0.1155  0.9123  0.6565 -0.8252  0.1070  0.3498  0.1888
#> 
#> Columns 19 to 20  0.5379 -0.3365
#>  -0.5983  0.7670
#>  -0.8661 -0.6540
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1259 -0.0482  0.2802  0.3610 -0.0992 -0.2024  0.6128 -0.2248  0.2926
#>  -0.2400  0.4402 -0.4100 -0.3112  0.0699  0.0050  0.2092  0.5696  0.4289
#>  -0.4690  0.4168 -0.4169 -0.3318  0.4766  0.0654  0.3764  0.5081  0.5479
#> 
#> Columns 10 to 18 -0.3072  0.0972  0.0485  0.6474  0.8042 -0.4972 -0.1274  0.5884  0.5483
#>   0.5106  0.3623  0.2485  0.0839 -0.6881 -0.6580  0.2933  0.0059 -0.0845
#>   0.0804  0.5548  0.2605 -0.2146 -0.4235 -0.2680  0.3610  0.4149 -0.0362
#> 
#> Columns 19 to 20 -0.7846 -0.0960
#>  -0.2463 -0.0071
#>   0.0809  0.2169
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2501  0.3705 -0.2369 -0.2074  0.6777  0.3847  0.5115  0.1592  0.5893
#>   0.0184  0.0467  0.1172  0.3934  0.2912  0.0785  0.1663 -0.1582 -0.3610
#>   0.1094 -0.1466 -0.1524  0.7878 -0.0610  0.4756 -0.0793  0.3487 -0.5730
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1667  0.3028 -0.4051  0.0334 -0.7885 -0.3942 -0.2663  0.4182 -0.1186
#>   0.6499 -0.0132 -0.1258  0.1727 -0.7370 -0.5296 -0.3386 -0.5099  0.6327
#>  -0.4051  0.1203 -0.2734 -0.2911  0.2403 -0.9137  0.6374 -0.5103 -0.0852
#> 
#> Columns 10 to 18 -0.1817 -0.0210 -0.0814  0.2225  0.4596 -0.2666  0.0286  0.4933 -0.5987
#>   0.0154  0.5337 -0.2097  0.4481  0.4872 -0.7832  0.0729  0.5673 -0.3852
#>   0.6670 -0.4071 -0.1533  0.5280  0.1574  0.2069  0.3102 -0.2173  0.5901
#> 
#> Columns 19 to 20 -0.2880  0.5239
#>  -0.1721  0.7158
#>   0.7030  0.0177
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0130  0.2348  0.4066  0.4901  0.2695 -0.0462  0.1069  0.4187 -0.3964
#>  -0.1412 -0.0895  0.0705  0.7327  0.1495  0.0296  0.1055  0.4110 -0.0168
#>  -0.0620 -0.1070 -0.4240  0.3966  0.5293 -0.1741  0.5519 -0.0118  0.7431
#> 
#> Columns 10 to 18 -0.0348  0.3144  0.2511  0.1360 -0.3082  0.1517  0.2628 -0.0839  0.3591
#>  -0.2166  0.1001  0.1961  0.3236  0.0482  0.0134  0.0376  0.0749  0.0261
#>  -0.3613 -0.3974 -0.4476  0.2675 -0.2790 -0.3638  0.5358 -0.0602  0.4789
#> 
#> Columns 19 to 20 -0.2938 -0.3506
#>  -0.4267 -0.3342
#>  -0.4460 -0.5558
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>