RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5933  0.8597 -0.3431 -0.8629  0.4803  0.8201  0.5044 -0.1441  0.6563
#>   0.8558  0.5000 -0.8850 -0.2523  0.3933 -0.7719  0.2013 -0.7454  0.2827
#>  -0.8868 -0.5842 -0.5476  0.8306  0.7731 -0.2176 -0.6723 -0.7194  0.9172
#> 
#> Columns 10 to 18  0.1568 -0.0343  0.0930  0.1584 -0.7794 -0.1000 -0.1496  0.5172  0.0483
#>  -0.2980 -0.5552  0.1549 -0.1017  0.9024 -0.3922  0.1687  0.5830 -0.5688
#>  -0.6057  0.8969  0.8017 -0.4543  0.1390  0.9298 -0.2615 -0.6703 -0.9059
#> 
#> Columns 19 to 20 -0.9190 -0.4300
#>  -0.4989  0.4632
#>  -0.7347  0.5230
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3226  0.6881 -0.1859  0.0945  0.4891 -0.0039 -0.1686 -0.4002 -0.2602
#>   0.0883  0.5781  0.2095  0.2301  0.1059  0.3221 -0.2534 -0.2771  0.5027
#>  -0.0925 -0.0557  0.4823 -0.0096  0.2670  0.2486 -0.3446  0.6473  0.7558
#> 
#> Columns 10 to 18 -0.2631 -0.1380 -0.0605 -0.0920  0.0423  0.5979 -0.0986 -0.2185  0.3761
#>   0.1911  0.6304 -0.2694  0.8786 -0.3778 -0.0374 -0.0786 -0.5229 -0.3005
#>  -0.8424  0.0637 -0.4552  0.2430 -0.2912 -0.1996 -0.1226 -0.7669  0.3022
#> 
#> Columns 19 to 20 -0.7465 -0.1133
#>   0.2644  0.1145
#>  -0.3577 -0.1170
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0455  0.4341 -0.0708  0.4101  0.3383 -0.0472 -0.4818 -0.3462 -0.1133
#>   0.0600  0.1337  0.6594 -0.3583  0.5768 -0.1967 -0.3756 -0.2931  0.3203
#>   0.1958  0.5370  0.0794  0.0295  0.6381 -0.1264 -0.0366 -0.1969  0.2434
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4414 -0.0745 -0.3141  0.4269 -0.1684 -0.5845  0.2812 -0.3449  0.6745
#>   0.4950 -0.4151  0.2503  0.4579  0.2517 -0.4699  0.8372  0.1097 -0.2155
#>   0.5015  0.0545  0.0273  0.7077 -0.1619 -0.3662  0.5139  0.0041 -0.1428
#> 
#> Columns 10 to 18 -0.6888 -0.8251  0.5915  0.4361  0.5090 -0.8295 -0.0171  0.0787  0.7646
#>  -0.6269 -0.5438  0.7354  0.6078  0.1581 -0.5984 -0.1787  0.5631  0.2318
#>   0.2935 -0.5779  0.5819  0.2608 -0.4472 -0.3137  0.2209 -0.3226  0.2657
#> 
#> Columns 19 to 20 -0.1712 -0.4818
#>  -0.0332 -0.0407
#>   0.0110 -0.1206
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0026  0.0113  0.1002 -0.0405  0.3671 -0.0545 -0.2876 -0.3005  0.3271
#>  -0.2401  0.6668 -0.1616 -0.0058  0.3935  0.0247 -0.1689 -0.4330  0.2134
#>  -0.1148  0.7052  0.2072 -0.2480  0.0910  0.3148 -0.1565 -0.2843 -0.0026
#> 
#> Columns 10 to 18 -0.6737  0.3307 -0.0850  0.0091  0.0724  0.1784  0.1500 -0.4608  0.0610
#>  -0.3411  0.3447 -0.3602  0.3461 -0.0633  0.2645  0.2385 -0.3613  0.2111
#>  -0.5205 -0.0169 -0.0524  0.2985  0.2428  0.4588  0.1391 -0.3030 -0.0187
#> 
#> Columns 19 to 20 -0.1745  0.2009
#>  -0.2498  0.0840
#>  -0.4370 -0.2270
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>