RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.7993 -0.7976  0.2385  0.3131 -0.7709  0.0121  0.4953  0.6468  0.7405
#>   0.4193  0.0146  0.3556 -0.4297 -0.6355  0.9591 -0.7063 -0.0919 -0.4081
#>   0.0368 -0.2607  0.6028  0.2835 -0.1843 -0.2228 -0.9056 -0.0647 -0.6504
#> 
#> Columns 10 to 18  0.8196 -0.8302  0.6831  0.1734  0.4629  0.6746 -0.6797 -0.6383 -0.3687
#>   0.5994 -0.1915  0.3532 -0.5729 -0.1049 -0.3749 -0.0410  0.2361  0.1198
#>   0.1174 -0.1293 -0.5502 -0.1363 -0.0883 -0.3146  0.8041 -0.7713 -0.6721
#> 
#> Columns 19 to 20  0.4891  0.5599
#>   0.0911  0.7512
#>   0.4750 -0.7471
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3636  0.0425 -0.0339 -0.0255  0.0827 -0.2331  0.2283 -0.5369  0.0668
#>   0.0618 -0.4323  0.0299 -0.0108 -0.2484  0.3574 -0.2208 -0.1541  0.5961
#>  -0.1155  0.0487 -0.0894  0.7275 -0.3157  0.0310 -0.6171  0.2208  0.2233
#> 
#> Columns 10 to 18  0.3373  0.3746  0.2317 -0.2138  0.4389  0.4272  0.5468  0.0489 -0.2285
#>   0.0814 -0.4535 -0.0228  0.1716  0.2087  0.3267 -0.6878 -0.4941 -0.5850
#>   0.2659 -0.0120  0.3336  0.3480  0.2752  0.2915 -0.5793 -0.1920 -0.6021
#> 
#> Columns 19 to 20 -0.4878  0.4254
#>   0.3174  0.2851
#>  -0.4374 -0.2472
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2976  0.0818 -0.0654 -0.3400  0.1905  0.2868  0.2807 -0.6326  0.1367
#>  -0.4965 -0.3789 -0.3607 -0.0044  0.1959  0.3216  0.1029 -0.3093  0.1406
#>  -0.1751 -0.0928 -0.2943  0.1128 -0.0648  0.0542  0.1675  0.0922  0.2310
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5231 -0.0902  0.2163 -0.3948  0.0475 -0.0300 -0.4014 -0.1334 -0.5242
#>  -0.3761  0.2714 -0.0293 -0.0640  0.4656 -0.0130  0.2974  0.0715  0.0801
#>  -0.2382  0.1420 -0.1526 -0.5467  0.3520 -0.1065 -0.7718 -0.4036  0.1062
#> 
#> Columns 10 to 18  0.6807 -0.4556  0.1791 -0.1758  0.6101  0.3980 -0.5085  0.0626  0.1813
#>  -0.0165  0.4801  0.7556  0.2085 -0.1023 -0.2479  0.5222  0.0760  0.2729
#>  -0.2487  0.4765  0.7078 -0.2761 -0.5012 -0.3239  0.5320  0.3372 -0.1100
#> 
#> Columns 19 to 20 -0.3433 -0.1115
#>  -0.5667  0.1369
#>  -0.5340  0.1031
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3401  0.3958  0.0303  0.3529 -0.2220  0.0332  0.0148 -0.3826  0.1057
#>  -0.3787 -0.1255 -0.2025 -0.2066  0.4484  0.1402  0.1805  0.0656  0.5532
#>   0.0087  0.2160 -0.2148  0.2663  0.2662  0.0185 -0.0440  0.2959  0.5819
#> 
#> Columns 10 to 18  0.0119  0.0204  0.2054  0.2476  0.3299  0.0368  0.1051 -0.6128 -0.2537
#>  -0.1570 -0.0404  0.2381  0.0519  0.2763  0.4020 -0.1485 -0.2639 -0.5741
#>   0.0253  0.1961  0.3658  0.0544  0.1710  0.5462  0.0316 -0.1336 -0.7031
#> 
#> Columns 19 to 20 -0.3597 -0.1245
#>   0.3834  0.0682
#>  -0.1013  0.3833
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>