RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3708  0.8419  0.9011 -0.4734  0.4108  0.5318 -0.2803  0.5298 -0.2342
#>   0.7469 -0.3697  0.3751  0.4613 -0.4781  0.4933  0.2897 -0.2724 -0.0074
#>   0.8906  0.1756 -0.4409  0.6318 -0.5761  0.6887 -0.7012 -0.5273 -0.7005
#> 
#> Columns 10 to 18  0.0322  0.1562  0.4492 -0.7635 -0.4543 -0.1564 -0.2496  0.4685 -0.2509
#>  -0.3605 -0.2295  0.4372  0.0171 -0.5289  0.1974  0.4579  0.3320 -0.4769
#>   0.0994  0.1977  0.7692 -0.2380  0.3436  0.7334  0.1746 -0.5712  0.6176
#> 
#> Columns 19 to 20  0.1797 -0.3747
#>  -0.2619  0.5395
#>   0.6015  0.5937
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2175  0.4014  0.2859 -0.0681 -0.2907  0.3671 -0.0261  0.4724  0.8514
#>   0.0044  0.4522  0.3944 -0.0189  0.2554  0.1392 -0.4243  0.0971  0.6056
#>  -0.6052  0.5352  0.0802 -0.6417 -0.4718  0.5714  0.2633  0.2198  0.7690
#> 
#> Columns 10 to 18 -0.4164  0.2350 -0.2300 -0.0700 -0.4489 -0.2685 -0.5377 -0.1412 -0.0598
#>  -0.2364  0.1707 -0.1603 -0.5118 -0.0552 -0.1800  0.2520  0.3842 -0.4414
#>   0.0915  0.6073 -0.8060 -0.4198  0.0203 -0.3846  0.1962 -0.6267 -0.4174
#> 
#> Columns 19 to 20  0.3555 -0.1183
#>  -0.0387 -0.0221
#>  -0.1147 -0.0445
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1417  0.2446  0.4793 -0.4817  0.2967  0.2028  0.5184  0.1116  0.2581
#>   0.0030  0.0649  0.2193 -0.2167  0.1804  0.0055 -0.2465 -0.0969  0.3914
#>  -0.2706 -0.0049  0.5618  0.0981  0.4868 -0.2481  0.5085 -0.4690  0.3209
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0040  0.3463 -0.3656 -0.2469  0.4445 -0.7066  0.4971 -0.1630  0.1794
#>   0.1735  0.4804 -0.3289  0.2851  0.4873 -0.0032  0.1858  0.5735  0.0756
#>  -0.2263 -0.4497 -0.0799 -0.8185 -0.0002 -0.6403 -0.6082 -0.6235 -0.1514
#> 
#> Columns 10 to 18  0.2983 -0.6159  0.2757  0.4288 -0.7705  0.5496 -0.0406 -0.3804  0.2028
#>  -0.5970 -0.0051 -0.0603 -0.1862  0.0417 -0.2500 -0.2358 -0.0642 -0.2110
#>  -0.0701 -0.6935 -0.2174 -0.2583 -0.8020 -0.5593 -0.1824  0.6529  0.3694
#> 
#> Columns 19 to 20 -0.5713  0.5480
#>   0.7315 -0.9257
#>  -0.2814 -0.3865
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3831  0.2752 -0.2922 -0.2127  0.0531  0.3318 -0.1752 -0.2366 -0.1663
#>  -0.0163 -0.0673  0.3095 -0.1535 -0.0701  0.1034 -0.0931  0.2735  0.4380
#>   0.2128 -0.3025 -0.1619  0.0359  0.0747  0.1338 -0.4913 -0.4422  0.5402
#> 
#> Columns 10 to 18 -0.3578  0.4038  0.4863  0.0067  0.0139 -0.0571 -0.1266 -0.2619  0.2263
#>  -0.5582  0.5901  0.0720  0.1214 -0.0582 -0.1740  0.1517 -0.1631 -0.3541
#>   0.1456  0.0007  0.5725 -0.1695 -0.0605 -0.4123 -0.0310  0.4452  0.0672
#> 
#> Columns 19 to 20 -0.0299  0.2853
#>  -0.4765  0.2095
#>  -0.3644  0.0686
#> [ CPUFloatType{2,3,20} ]
#>