RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1848  0.6946 -0.0462 -0.7394 -0.6843 -0.5613  0.8171 -0.6003  0.6926
#>   0.6386  0.3294  0.1016 -0.3576 -0.8103  0.2339  0.7611 -0.3209  0.0531
#>   0.0088  0.3854 -0.1836 -0.0011 -0.1497  0.3847  0.1563 -0.7619  0.2597
#> 
#> Columns 10 to 18  0.5512 -0.0486 -0.1662 -0.3033 -0.5070  0.2406  0.3643 -0.5288 -0.5947
#>  -0.7424 -0.1120 -0.8340  0.6208 -0.6086  0.0305 -0.2825  0.8372  0.0731
#>   0.7027 -0.0522  0.3691 -0.4330 -0.0709 -0.2209 -0.7872  0.0036 -0.1052
#> 
#> Columns 19 to 20  0.5240  0.7612
#>   0.6259  0.8903
#>   0.0693 -0.0548
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4933  0.8052 -0.1079 -0.2769 -0.3266  0.4516  0.0442  0.4266  0.4961
#>  -0.4199  0.8168  0.2387 -0.1815 -0.5811 -0.1661  0.7419  0.4456 -0.0703
#>   0.1174  0.1466 -0.0730  0.0174 -0.3708 -0.1597  0.2325 -0.1157  0.5659
#> 
#> Columns 10 to 18  0.2067 -0.3317 -0.3782 -0.3383  0.4151  0.3542 -0.4190  0.3514 -0.6307
#>   0.5065 -0.3978 -0.4729 -0.3486  0.3073  0.3370  0.1818 -0.1840 -0.3775
#>   0.5627 -0.2604 -0.3222 -0.3881  0.3888  0.1626 -0.4054 -0.0960  0.3677
#> 
#> Columns 19 to 20 -0.0319  0.4762
#>   0.0070  0.3102
#>   0.6090  0.6362
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4091  0.2563  0.1318  0.1484 -0.5077 -0.2156  0.1068 -0.0388  0.6234
#>   0.0903  0.2479  0.2111  0.1445 -0.4068 -0.0140  0.0095  0.1413  0.5617
#>   0.2436  0.4455  0.2080 -0.1787 -0.6533 -0.0963  0.0731  0.1794  0.4325
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0179 -0.1922  0.3359  0.0231 -0.0763 -0.2587  0.6072 -0.1747 -0.8002
#>   0.5020 -0.1131  0.3013  0.5015  0.5695  0.0239 -0.1651  0.2500 -0.2512
#>  -0.4077 -0.8326 -0.2657  0.1825 -0.0242 -0.4975 -0.0002  0.0698 -0.1704
#> 
#> Columns 10 to 18  0.9504 -0.7580  0.2743 -0.5311  0.1822  0.2131  0.2766  0.2483  0.4832
#>  -0.1770 -0.7813  0.6740  0.7114 -0.7339 -0.1078  0.3927  0.3722 -0.7951
#>   0.4549  0.4910 -0.0466  0.0738 -0.1916  0.1320  0.6943  0.0045  0.5781
#> 
#> Columns 19 to 20 -0.0070 -0.2138
#>  -0.3348 -0.0427
#>  -0.6685  0.0746
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0839  0.2236  0.2145  0.0120 -0.2250 -0.0872  0.0328  0.1023  0.3371
#>   0.4785  0.4394  0.2312  0.0583 -0.0707 -0.1698  0.2807  0.1765  0.1251
#>  -0.0811  0.3504  0.2070  0.1879 -0.3807 -0.1525  0.0327 -0.1341  0.6405
#> 
#> Columns 10 to 18  0.4189 -0.2899 -0.6863  0.0086 -0.0478  0.0901 -0.0573  0.4198  0.0603
#>   0.5719 -0.4135  0.1526  0.2537  0.3715 -0.0696  0.0751 -0.0909  0.1069
#>   0.1672 -0.0069 -0.5971  0.2431  0.2819  0.0757 -0.0716  0.2165 -0.2789
#> 
#> Columns 19 to 20  0.1590  0.5068
#>   0.5313  0.8622
#>   0.0731 -0.1200
#> [ CPUFloatType{2,3,20} ]
#>