RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0929 -0.2474 -0.3517 -0.1624  0.3695 -0.7993 -0.1019 -0.4718 -0.3383
#>   0.9491 -0.7629  0.9149  0.7603  0.0351 -0.7882 -0.6068 -0.6490 -0.5146
#>   0.7024 -0.4617  0.5516  0.5359  0.4448 -0.2315 -0.9654 -0.6095 -0.4421
#> 
#> Columns 10 to 18  0.3113  0.1941  0.7945  0.3346  0.2722  0.3395 -0.3014 -0.5461  0.2949
#>   0.3597 -0.8241 -0.3636  0.7277 -0.3620 -0.6150 -0.5482 -0.0953  0.9006
#>  -0.9639 -0.9896 -0.0596  0.6913 -0.3694  0.6003  0.6927  0.2548  0.2921
#> 
#> Columns 19 to 20 -0.0099  0.8035
#>  -0.2837  0.9349
#>  -0.4249  0.4073
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0895  0.0017  0.1801 -0.4017  0.5400  0.0487 -0.2656  0.5274  0.5395
#>   0.3694  0.2207  0.3306  0.4001  0.1436  0.0630 -0.9201  0.0686 -0.0646
#>   0.4199 -0.1598 -0.3105  0.4765  0.2727  0.3100 -0.4955  0.3185  0.1752
#> 
#> Columns 10 to 18  0.1061  0.7065  0.6023  0.3735  0.0987 -0.0652 -0.1373 -0.0497  0.4106
#>  -0.3232  0.1346  0.1209 -0.1192  0.2954  0.3521 -0.1422  0.6128 -0.0793
#>   0.0371 -0.0207  0.4459  0.1416  0.4776  0.1157  0.1626  0.6663  0.1389
#> 
#> Columns 19 to 20  0.0488  0.5369
#>  -0.4861  0.6649
#>  -0.5258  0.4923
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1208 -0.2079 -0.4222 -0.3076  0.4068 -0.1872 -0.5268 -0.1999  0.1646
#>  -0.0250 -0.3927  0.1388  0.2074  0.3291  0.4687 -0.4392  0.1795 -0.1897
#>  -0.1053 -0.2943 -0.0666  0.4830  0.6171  0.2256 -0.2695 -0.0811 -0.5247
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0577 -0.0285 -0.6097 -0.4585 -0.5266 -0.5030  0.4769 -0.1149  0.4807
#>  -0.1984  0.2439  0.0649  0.0664 -0.2871 -0.4076 -0.0557 -0.5437  0.4028
#>  -0.1435 -0.5155 -0.5990 -0.2413 -0.1453  0.6784 -0.3300 -0.5952 -0.8189
#> 
#> Columns 10 to 18  0.5801  0.5575 -0.1683  0.2659  0.8862 -0.1161 -0.4274  0.1793 -0.0401
#>   0.2667  0.5896  0.6217 -0.0437  0.3957 -0.3869 -0.5589  0.0864  0.0765
#>   0.0643 -0.5095 -0.7458 -0.1369 -0.7916  0.3587 -0.0986  0.6167  0.8582
#> 
#> Columns 19 to 20 -0.1983  0.2667
#>   0.5927  0.1416
#>   0.5483 -0.4940
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2980 -0.3391 -0.3825 -0.1073  0.2459  0.1564 -0.0540 -0.2768  0.3198
#>   0.2353 -0.5407 -0.0084  0.2127  0.1943  0.1406 -0.2141 -0.3979  0.2578
#>   0.3298 -0.2961  0.2067  0.6027  0.1831  0.5840 -0.4362  0.1327 -0.3851
#> 
#> Columns 10 to 18 -0.4043  0.0306  0.2919  0.1249  0.1170  0.0134  0.3242  0.5945  0.5805
#>   0.1099 -0.1660  0.3703 -0.0664  0.0664 -0.2685 -0.1187  0.4981  0.7128
#>  -0.3554 -0.2718  0.2618  0.1066  0.5388 -0.1473 -0.0853  0.4555 -0.0724
#> 
#> Columns 19 to 20  0.0690  0.3524
#>  -0.0041  0.3597
#>   0.0435  0.1076
#> [ CPUFloatType{2,3,20} ]
#>