RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2961 -0.0779 -0.5222  0.3360 -0.4824  0.0817 -0.6743  0.9239  0.9133
#>   0.0604  0.6535 -0.8408  0.0444 -0.8183  0.8535 -0.6456  0.2660  0.5218
#>  -0.1815 -0.6878  0.0010 -0.8232  0.8907  0.1331 -0.3337  0.3851  0.3290
#> 
#> Columns 10 to 18 -0.8731  0.7373  0.5978  0.1651 -0.8215 -0.5048 -0.3064  0.2003 -0.6331
#>  -0.5618  0.3141  0.1819  0.2553 -0.3283 -0.3272 -0.3945 -0.1736 -0.6469
#>  -0.0293  0.6194  0.5829  0.5448  0.7681 -0.2698 -0.6569  0.1816 -0.6972
#> 
#> Columns 19 to 20 -0.0432 -0.7985
#>   0.3342 -0.4597
#>   0.0351  0.8292
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4998  0.2951 -0.7133 -0.2840 -0.4322 -0.0213  0.2669  0.4181  0.1530
#>   0.2935  0.2186 -0.1974  0.0459 -0.0552 -0.4372  0.4473  0.2559  0.3964
#>  -0.3680 -0.0477 -0.5290 -0.1750 -0.3808 -0.1329  0.2055 -0.4607  0.1036
#> 
#> Columns 10 to 18 -0.1875  0.3758  0.1121 -0.7110 -0.3435 -0.1054 -0.0946 -0.5469 -0.2348
#>   0.4302  0.2218  0.2837 -0.4271 -0.3085 -0.3444  0.1394 -0.1500  0.4131
#>  -0.5546  0.3785  0.3211 -0.2688 -0.5751  0.2965 -0.7247 -0.5031  0.2525
#> 
#> Columns 19 to 20  0.4327 -0.6341
#>   0.4680 -0.5836
#>   0.3808  0.1956
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2217 -0.2558 -0.5934  0.1449 -0.1623 -0.2749  0.2270 -0.2782 -0.2870
#>   0.5056 -0.2198 -0.3801  0.1769  0.3384 -0.1689  0.0787 -0.4227 -0.1139
#>  -0.0335 -0.1235 -0.3718 -0.3211 -0.2201 -0.2724 -0.3428  0.1100 -0.1174
#> ... [the output was truncated (use n=-1 to disable)]
#>  -0.0111 -0.1493
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3057 -0.1561 -0.2747 -0.0534 -0.2968  0.5209 -0.6099 -0.4368  0.5645
#>  -0.3277 -0.4267  0.1691 -0.6021  0.3849  0.1173  0.8366  0.3452 -0.5994
#>  -0.3478  0.6728 -0.5875 -0.0788  0.3853 -0.1031  0.0071 -0.3602 -0.6338
#> 
#> Columns 10 to 18  0.0045  0.4373  0.5175  0.4374  0.4860 -0.7337  0.5910  0.8039 -0.2163
#>   0.3261  0.2156  0.7942  0.0795  0.3229  0.0500  0.4462 -0.3757  0.6231
#>   0.5994  0.6461  0.6742  0.4427  0.4273 -0.0356  0.2896  0.3971  0.0029
#> 
#> Columns 19 to 20  0.0809 -0.1132
#>  -0.7680  0.7701
#>   0.3080 -0.1482
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0333  0.0265 -0.4772  0.2658 -0.3959 -0.0625  0.0375 -0.2778  0.1312
#>   0.0966  0.2974 -0.5540 -0.4692 -0.3315  0.0937 -0.3309  0.1800  0.5699
#>  -0.3020  0.0493 -0.0766  0.0917 -0.4885 -0.1259  0.0521 -0.1315 -0.0320
#> 
#> Columns 10 to 18 -0.3694  0.6226  0.1839 -0.1325 -0.2366 -0.1387  0.1830 -0.1448 -0.1248
#>  -0.2794  0.5545  0.3320  0.0762 -0.2031 -0.0258 -0.3559 -0.3180 -0.2741
#>  -0.1158  0.1363  0.3174  0.2896  0.0134 -0.5935  0.0443  0.3625 -0.1952
#> 
#> Columns 19 to 20  0.2359 -0.4570
#>   0.0057 -0.2850
#>  -0.0111 -0.1493
#>