RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2773  0.5561  0.3568 -0.1149 -0.1855 -0.6522  0.1852  0.3647  0.7883
#>   0.0162 -0.8259  0.7485  0.6966  0.1757 -0.8638  0.8103 -0.5602 -0.5368
#>   0.0984  0.6862  0.4927 -0.7208  0.1802  0.1237 -0.0887  0.2157  0.1663
#> 
#> Columns 10 to 18  0.6027  0.2500 -0.4384  0.2614  0.1153 -0.6572  0.2664  0.3201 -0.6423
#>   0.0116  0.5602  0.6607  0.0048  0.3744  0.2898 -0.4793  0.4391 -0.6568
#>  -0.0839  0.4259 -0.2150 -0.5696  0.3213 -0.0787 -0.4651  0.4115  0.2776
#> 
#> Columns 19 to 20  0.5123 -0.1510
#>  -0.1544 -0.6348
#>  -0.5303  0.7005
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2452  0.1291 -0.3821 -0.5715  0.3413 -0.1938 -0.3640  0.2483 -0.0663
#>   0.3744 -0.5220 -0.1820  0.6239  0.2414 -0.3296  0.7707 -0.4365 -0.2631
#>   0.5646 -0.2297  0.5355 -0.0920  0.6885  0.0769 -0.0498 -0.4345 -0.4209
#> 
#> Columns 10 to 18 -0.6320  0.3118  0.4271 -0.2328 -0.2529 -0.4808 -0.1340  0.3306 -0.6244
#>   0.2095  0.4414  0.2887  0.5890 -0.0988  0.3440 -0.2142  0.1546  0.6159
#>  -0.5150  0.4801  0.4823 -0.6042  0.1675 -0.0502 -0.5160 -0.0200 -0.0165
#> 
#> Columns 19 to 20  0.4853 -0.4303
#>  -0.7011  0.2811
#>  -0.4708 -0.3491
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5561 -0.0587 -0.0817 -0.4912  0.6281 -0.2573 -0.0994 -0.2927 -0.3671
#>   0.0848 -0.0082  0.0195  0.4354  0.1204 -0.1273  0.2144 -0.5420  0.0685
#>   0.2041 -0.1208  0.1244  0.1852  0.2865 -0.2526  0.6022 -0.5774 -0.0898
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6117 -0.5767  0.0528  0.8930  0.0539  0.3929  0.0744 -0.7072  0.2727
#>  -0.0358  0.0737  0.4805  0.4315  0.3592 -0.4708  0.0924  0.1556 -0.0639
#>  -0.1951 -0.1987  0.7100  0.5443  0.4522 -0.7765 -0.2148  0.5473  0.0509
#> 
#> Columns 10 to 18  0.5136 -0.4458 -0.3934  0.2728 -0.6839  0.0655  0.4676  0.1340  0.0364
#>   0.1498 -0.5516 -0.6641 -0.0154  0.2493  0.8494  0.3503  0.6599 -0.0634
#>  -0.0030 -0.2363 -0.2158  0.1860 -0.3722  0.2836 -0.1202  0.4673  0.0345
#> 
#> Columns 19 to 20  0.7165 -0.8693
#>   0.6075 -0.4945
#>   0.3546 -0.7570
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.3008  0.5299  0.3223 -0.4260 -0.2648 -0.1564 -0.1954  0.0077  0.1552
#>   0.0504  0.1948 -0.1301 -0.1826 -0.0118 -0.1477  0.1734  0.0514  0.2495
#>   0.2449  0.2061  0.0712  0.1921 -0.0700 -0.0292 -0.0587 -0.0298 -0.1516
#> 
#> Columns 10 to 18 -0.1536  0.2115  0.6027 -0.7733  0.8466 -0.6495 -0.0588 -0.0683 -0.2472
#>  -0.3879  0.2557  0.6413 -0.0866  0.2697 -0.3826 -0.3066 -0.2901 -0.2312
#>  -0.2961  0.3722  0.7382 -0.2280  0.4259 -0.1635 -0.4625 -0.2205  0.1002
#> 
#> Columns 19 to 20  0.0543 -0.1945
#>  -0.0533 -0.3463
#>  -0.3295  0.0223
#> [ CPUFloatType{2,3,20} ]
#>