RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5458  0.6804  0.4956 -0.2836  0.6698  0.4193 -0.3768 -0.4724 -0.9321
#>   0.3719 -0.1947  0.2253  0.0482 -0.6605  0.0141 -0.8970  0.0958  0.1226
#>  -0.3970  0.4632  0.1164  0.0599  0.0374 -0.3183  0.9272 -0.5680 -0.5253
#> 
#> Columns 10 to 18 -0.3395 -0.7413 -0.6656  0.6796  0.7662  0.7525 -0.6108  0.2986  0.1671
#>  -0.0097 -0.5455  0.6649  0.4684 -0.1372  0.6427  0.7471 -0.7244  0.4196
#>   0.7363 -0.7374 -0.4804 -0.2095  0.3546 -0.1049 -0.9024  0.2194 -0.0768
#> 
#> Columns 19 to 20  0.5900  0.7828
#>  -0.1941 -0.1871
#>   0.8516 -0.1118
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1591  0.0441 -0.0060 -0.4413  0.5876 -0.6077  0.4828  0.2144 -0.3169
#>   0.4436  0.4612  0.1947 -0.0964 -0.5168 -0.1081  0.0242 -0.2661  0.0746
#>   0.4926 -0.0898  0.3205 -0.3512  0.6337 -0.2414 -0.1743  0.4797 -0.4480
#> 
#> Columns 10 to 18 -0.2461 -0.3987  0.5667  0.0290  0.5919 -0.5859 -0.1145 -0.1326  0.1305
#>   0.6370 -0.0865  0.2964  0.5509 -0.1240  0.0500  0.5244 -0.3490  0.3970
#>  -0.0759 -0.0133  0.3313  0.0569 -0.0284 -0.6680 -0.7094  0.6971 -0.5205
#> 
#> Columns 19 to 20  0.4888 -0.5028
#>   0.3064  0.1523
#>   0.2046 -0.2503
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.5977  0.2805  0.4481 -0.3144  0.2655 -0.0816 -0.1918 -0.1043 -0.4049
#>   0.3620  0.2452  0.1509 -0.5182  0.3793  0.0971 -0.4544  0.0669  0.2815
#>   0.6103  0.1907  0.2446 -0.1172  0.2068  0.1459  0.3303 -0.3699 -0.3576
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5545 -0.5758 -0.6046  0.1521  0.3966 -0.8106  0.2225 -0.0490  0.3410
#>  -0.4224 -0.7572 -0.1976  0.4609  0.4332  0.0401  0.0383  0.0813  0.2133
#>  -0.7564  0.0603 -0.0922  0.2833 -0.3898  0.2026 -0.3203 -0.0562  0.0176
#> 
#> Columns 10 to 18 -0.1139 -0.0594  0.2201  0.1304  0.4575  0.3626 -0.3294 -0.1381 -0.0560
#>  -0.0334 -0.6083  0.0140 -0.1038  0.3824 -0.2782 -0.1764 -0.0741  0.0791
#>  -0.2587 -0.4729 -0.5345 -0.3862 -0.8314 -0.7488 -0.0135 -0.6276 -0.8680
#> 
#> Columns 19 to 20 -0.1850 -0.3379
#>   0.1270 -0.5008
#>   0.2848 -0.7906
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4751  0.2119 -0.0521  0.0598 -0.0180  0.0458 -0.4223 -0.0833 -0.0784
#>   0.5306  0.2734 -0.0481  0.2045 -0.1390 -0.1322 -0.1303  0.0675 -0.2539
#>   0.5032  0.1371  0.1129 -0.5964  0.3235  0.1235 -0.2225 -0.2124 -0.0375
#> 
#> Columns 10 to 18  0.2073 -0.2525  0.4475 -0.1307  0.4278  0.0551  0.5993 -0.0862  0.3277
#>   0.4418 -0.5563  0.1231  0.4083  0.2161  0.0611  0.2308 -0.0392  0.5531
#>  -0.0114 -0.4719  0.3878  0.0121  0.5930 -0.0083 -0.4512 -0.2651 -0.1457
#> 
#> Columns 19 to 20  0.7016  0.5645
#>   0.6088  0.3438
#>   0.3923  0.0235
#> [ CPUFloatType{2,3,20} ]
#>