RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3633  0.0453  0.6207 -0.5088 -0.1411  0.2220  0.3593  0.6115 -0.0718
#>  -0.6233 -0.5135 -0.2795 -0.5211 -0.6284  0.6404 -0.1081  0.1755  0.5666
#>  -0.3649  0.5090 -0.1395  0.5014 -0.1942  0.1873  0.2639  0.0293 -0.0218
#> 
#> Columns 10 to 18 -0.5941 -0.5389  0.0626  0.7607  0.3983  0.6433 -0.0352  0.3639  0.6408
#>  -0.7307 -0.3213 -0.5802  0.0854  0.0710 -0.3993  0.8480 -0.0822  0.8700
#>  -0.2732  0.1192  0.1923 -0.6551  0.1472  0.2335  0.1220  0.8031 -0.0921
#> 
#> Columns 19 to 20  0.3526  0.6823
#>  -0.8676 -0.7204
#>  -0.1518  0.8058
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0251  0.2838  0.0191 -0.7072  0.0976  0.4263 -0.2161 -0.3067 -0.5059
#>  -0.3106  0.0872  0.6565 -0.7406 -0.1178 -0.1533 -0.0424  0.1070 -0.0956
#>   0.1206  0.3879  0.0448 -0.0819  0.3359  0.0174 -0.4422  0.1508 -0.2833
#> 
#> Columns 10 to 18 -0.5530  0.3712 -0.3683 -0.1972  0.2142 -0.0407  0.2254 -0.3194  0.5658
#>   0.2344  0.2939 -0.6897  0.7683 -0.1582  0.1664  0.1123  0.0851 -0.0637
#>  -0.0286  0.1697 -0.4799  0.3162  0.0368  0.1915  0.2770 -0.0638  0.4738
#> 
#> Columns 19 to 20 -0.0651 -0.6150
#>   0.0129 -0.5666
#>  -0.1177 -0.0133
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4541  0.1251  0.0656 -0.4974 -0.3496 -0.0526  0.1163  0.2530  0.1231
#>  -0.3496 -0.2840 -0.1420 -0.7871 -0.2389  0.1727  0.0204 -0.2260 -0.1214
#>  -0.1897 -0.1302 -0.0250 -0.4952  0.1216  0.3290  0.0484 -0.1075 -0.1678
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.2469 -0.2026 -0.5074  0.0280 -0.4838 -0.3055  0.5529 -0.5215 -0.1547
#>   0.6377  0.0102  0.3513  0.0748  0.6600  0.2894 -0.7614 -0.0309  0.1460
#>   0.1660 -0.1941 -0.0717 -0.1674  0.2135  0.3850  0.0094 -0.1351 -0.2725
#> 
#> Columns 10 to 18 -0.4017  0.0234  0.3692 -0.3943  0.1645  0.2475  0.4328 -0.2772 -0.3723
#>   0.6401  0.4682  0.3653 -0.7590  0.1228  0.1963  0.6003 -0.5113 -0.2414
#>   0.3436  0.2804  0.2756  0.0880 -0.0051  0.1615  0.3986  0.2228  0.2104
#> 
#> Columns 19 to 20  0.3675 -0.3422
#>   0.1895  0.3232
#>   0.2451  0.5288
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1184  0.2062  0.1000 -0.0358  0.0141  0.4499  0.3516 -0.0694 -0.0470
#>   0.0005 -0.1053 -0.1303 -0.5556  0.1250  0.1398 -0.3944 -0.1523 -0.1786
#>  -0.0382 -0.0011  0.1037 -0.2084 -0.0369  0.3368 -0.1601 -0.0147  0.0900
#> 
#> Columns 10 to 18 -0.1082  0.4270  0.0632 -0.0471 -0.0971 -0.4344  0.3410  0.3321  0.4194
#>   0.0284  0.4221 -0.6225  0.2780 -0.2305  0.0159  0.5749 -0.5989  0.1410
#>  -0.1723 -0.0111 -0.1807  0.1180 -0.1597  0.0854  0.5220  0.2178  0.2016
#> 
#> Columns 19 to 20  0.0103 -0.2573
#>   0.3838 -0.2289
#>   0.2113  0.0524
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>