RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.2265 -0.6262  0.3592  0.0705 -0.0281 -0.5593  0.3801  0.5091  0.4315
#>   0.6305 -0.4957  0.7561  0.6756  0.0843  0.6375  0.6554  0.4797  0.0242
#>   0.2722  0.5789  0.4241 -0.3787  0.1893  0.4979  0.2005  0.6154 -0.6761
#> 
#> Columns 10 to 18 -0.7857  0.6342 -0.4320  0.7947  0.2046  0.2219  0.6459 -0.3636  0.2288
#>  -0.7197 -0.3396  0.5055 -0.4030  0.1038  0.4408  0.7378 -0.1914 -0.8637
#>   0.5893 -0.1425  0.5450  0.0930 -0.2945  0.2812 -0.3372 -0.1054 -0.8257
#> 
#> Columns 19 to 20  0.8456 -0.6223
#>   0.0725  0.6560
#>  -0.4331 -0.5673
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2070 -0.2139  0.8676  0.3823  0.4081  0.2198  0.4309  0.2087  0.2936
#>   0.4102  0.2110  0.1474  0.6109  0.3561  0.0192 -0.0810  0.2950 -0.3147
#>  -0.0339  0.3339  0.2479  0.1410 -0.6278 -0.1931 -0.0099  0.8177 -0.4417
#> 
#> Columns 10 to 18 -0.4397  0.0689 -0.3573 -0.1192  0.3013  0.1921  0.4794 -0.1974 -0.1238
#>  -0.6001 -0.5885 -0.2558  0.3035  0.0872  0.0426  0.0909 -0.3703 -0.0465
#>  -0.2286 -0.2619  0.4361 -0.0294 -0.0078 -0.0629 -0.1402  0.0090 -0.1049
#> 
#> Columns 19 to 20  0.3625 -0.0283
#>   0.7162 -0.2171
#>   0.2920 -0.1414
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0877  0.0670  0.2923  0.5123  0.3992  0.2602 -0.0789  0.0575 -0.1578
#>   0.0427  0.0424  0.2888  0.0333  0.2407  0.1183  0.5912 -0.1079 -0.1382
#>   0.0114  0.1693  0.3194  0.2406  0.1298  0.2739  0.2951  0.5354 -0.5762
#> 
#> Columns 10 to 18 -0.5458 -0.4453 -0.5065 -0.3118  0.2619 -0.0612  0.3299 -0.1688  0.1074
#>  -0.4059 -0.0766 -0.0907 -0.3775 -0.3504 -0.0091  0.4115  0.3085 -0.4517
#>  -0.2520 -0.4147 -0.1229 -0.1464  0.0490 -0.1875  0.1736 -0.3137 -0.0542
#> 
#> Columns 19 to 20  0.5254 -0.1976
#>   0.7996 -0.2098
#>   0.6788 -0.3888
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.0012  0.0495  0.5142 -0.0373 -0.1275  0.3090  0.6411  0.2349 -0.3433
#>   0.3434 -0.0446  0.4099  0.1868 -0.0692  0.0827  0.4273  0.2280 -0.3387
#>   0.1830 -0.0539  0.3358 -0.0435  0.0207  0.5368  0.1583  0.4330 -0.3587
#> 
#> Columns 10 to 18 -0.3289 -0.1215 -0.3483 -0.2492  0.0122 -0.0210  0.1180  0.0832 -0.3901
#>  -0.3096  0.2425 -0.1381 -0.5303  0.0377  0.3359  0.2373  0.1155  0.0341
#>  -0.5662 -0.4937 -0.4332 -0.2078  0.1073 -0.1748  0.2032 -0.5734 -0.2372
#> 
#> Columns 19 to 20  0.4491 -0.2495
#>   0.6905  0.0602
#>   0.4114 -0.1681
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.6483  0.1726  0.3433  0.1447 -0.3476  0.5816  0.4603  0.4855  0.0372
#>   0.4031  0.1653  0.2680 -0.2444  0.2267  0.3721 -0.0042  0.3200 -0.2639
#>   0.4371  0.1558  0.4998  0.2726  0.1165  0.3012  0.1602  0.1041 -0.1988
#> 
#> Columns 10 to 18 -0.2950 -0.3839 -0.2996 -0.4021  0.1674 -0.2379  0.1819 -0.2546 -0.6230
#>  -0.4707 -0.3934 -0.3938 -0.4774 -0.2556 -0.0530  0.3833 -0.4750 -0.5529
#>  -0.6916 -0.6371 -0.3869  0.1438 -0.0650 -0.2643  0.0851 -0.4978 -0.4724
#> 
#> Columns 19 to 20  0.2480 -0.0385
#>   0.4894 -0.4461
#>   0.4761 -0.3763
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4509  0.0518 -0.6660  0.5473 -0.1160  0.6275 -0.2288  0.0742 -0.6822
#>  -0.6383 -0.7364 -0.1665  0.4589 -0.4969  0.6828 -0.5775  0.3369 -0.5463
#>  -0.5647 -0.4898 -0.4073  0.8025  0.0386  0.4112 -0.7067  0.0587 -0.3376
#> 
#> Columns 10 to 18  0.0439 -0.5525 -0.7922  0.3833 -0.6046 -0.7473 -0.3672  0.7052 -0.1773
#>   0.3484 -0.3618  0.4232  0.4414  0.0141 -0.6755  0.0908  0.2657  0.1803
#>   0.7068 -0.3265 -0.1402 -0.2467  0.3635 -0.3478  0.0999  0.6179 -0.4193
#> 
#> Columns 19 to 20 -0.4612 -0.2237
#>   0.0611  0.4871
#>   0.4155  0.3709
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6483  0.1726  0.3433  0.1447 -0.3476  0.5816  0.4603  0.4855  0.0372
#>   0.4031  0.1653  0.2680 -0.2444  0.2267  0.3721 -0.0042  0.3200 -0.2639
#>   0.4371  0.1558  0.4998  0.2726  0.1165  0.3012  0.1602  0.1041 -0.1988
#> 
#> Columns 10 to 18 -0.2950 -0.3839 -0.2996 -0.4021  0.1674 -0.2379  0.1819 -0.2546 -0.6230
#>  -0.4707 -0.3934 -0.3938 -0.4774 -0.2556 -0.0530  0.3833 -0.4750 -0.5529
#>  -0.6916 -0.6371 -0.3869  0.1438 -0.0650 -0.2643  0.0851 -0.4978 -0.4724
#> 
#> Columns 19 to 20  0.2480 -0.0385
#>   0.4894 -0.4461
#>   0.4761 -0.3763
#> [ CPUFloatType{2,3,20} ]
#>