RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.7637 -0.1313  0.5969  0.1531 -0.1254  0.4891  0.1234  0.6253 -0.7492
#>   0.2589  0.4242  0.3661  0.1423 -0.0849  0.6521  0.6460 -0.0700  0.4534
#>   0.3733  0.6742 -0.2217  0.7105 -0.9161  0.3648  0.7792  0.9735  0.2612
#> 
#> Columns 10 to 18 -0.5984 -0.3374  0.4611 -0.8359  0.0462 -0.4568 -0.4345 -0.3505 -0.1522
#>   0.1463  0.7984 -0.3475 -0.4491 -0.0271  0.8812  0.8778  0.1531 -0.7390
#>   0.3041  0.4292 -0.0406  0.4252  0.0145  0.7781  0.7769  0.0593 -0.9202
#> 
#> Columns 19 to 20  0.3538  0.6781
#>  -0.0433 -0.2945
#>   0.6856 -0.6057
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6710 -0.5311  0.2803  0.0005  0.0275  0.3838  0.6671 -0.0162 -0.5996
#>   0.4107  0.5842  0.2051  0.3702 -0.0769  0.0561 -0.0541  0.7217 -0.0894
#>   0.4672  0.6836  0.2409  0.2795 -0.5998  0.5512  0.5058  0.7695 -0.0086
#> 
#> Columns 10 to 18 -0.0551 -0.2178 -0.0659  0.1676 -0.0844 -0.0615 -0.5221 -0.7109 -0.0620
#>  -0.4274 -0.2489 -0.4607 -0.1075  0.5351  0.4126  0.5633 -0.0264 -0.4220
#>  -0.3240 -0.0220 -0.1087  0.1033  0.6503  0.3712  0.1714  0.1283 -0.0168
#> 
#> Columns 19 to 20  0.3535  0.0174
#>   0.6221  0.1817
#>   0.3590  0.3580
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2701 -0.2973  0.4441  0.2044  0.1914  0.2487 -0.0745  0.3839 -0.3540
#>   0.2780  0.3469  0.2023  0.1925 -0.0489  0.6147 -0.1428  0.3720  0.0832
#>   0.2282  0.4958  0.4267  0.7054 -0.2349  0.7568  0.0074  0.8540  0.0266
#> 
#> Columns 10 to 18 -0.6464 -0.7034 -0.6523 -0.0177  0.0464 -0.4234  0.2099 -0.3220 -0.3344
#>  -0.5478 -0.2617  0.0352 -0.1966  0.0788  0.0812 -0.0471 -0.2582 -0.1233
#>  -0.3197 -0.2086 -0.3449  0.2758  0.5648  0.0263 -0.0207 -0.3915  0.3199
#> 
#> Columns 19 to 20  0.3531 -0.2637
#>   0.5581 -0.4702
#>   0.6524  0.3488
#> 
#> (4,.,.) = 
#>  Columns 1 to 9  0.5701 -0.1065  0.5072  0.1443 -0.0941  0.3671  0.8441 -0.0936 -0.1974
#>   0.4748  0.1017  0.4614  0.2620 -0.3558  0.2983  0.1250  0.6880 -0.3447
#>   0.3444  0.3684  0.1801  0.6321 -0.0474  0.3748  0.3065  0.0723  0.1163
#> 
#> Columns 10 to 18 -0.0605 -0.3080 -0.3329  0.1038 -0.0467 -0.1445  0.2260 -0.1848 -0.5946
#>  -0.1084  0.0672 -0.4801  0.0892  0.3854 -0.0134 -0.1895 -0.4964  0.0992
#>  -0.2141 -0.1909 -0.1810  0.0633  0.2411 -0.2714 -0.5017 -0.3317  0.3930
#> 
#> Columns 19 to 20  0.1336 -0.0267
#>   0.5617 -0.0372
#>   0.6570  0.2231
#> 
#> (5,.,.) = 
#>  Columns 1 to 9  0.4946  0.4707  0.2327  0.2931 -0.3486  0.2573  0.2574  0.6971 -0.1110
#>   0.4681  0.2777  0.3355  0.1704  0.0008  0.3576  0.1676  0.5053 -0.1168
#>   0.2939  0.3154  0.4486  0.3067 -0.0420  0.4443  0.4074  0.3560 -0.0950
#> 
#> Columns 10 to 18 -0.6061 -0.3754  0.0054  0.1500  0.2937 -0.2683  0.3698  0.0380 -0.3692
#>  -0.3798 -0.3725 -0.4083  0.2269  0.2623 -0.2974 -0.0191 -0.5680 -0.1990
#>  -0.3157 -0.0189 -0.4742 -0.3299  0.1420 -0.2435 -0.0617 -0.2264  0.1495
#> 
#> Columns 19 to 20  0.7276 -0.0975
#>   0.5028  0.0743
#>   0.7958  0.2401
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor 
#> (1,.,.) = 
#>  Columns 1 to 9  0.0724  0.0496 -0.7117  0.8362 -0.1116  0.4193 -0.2727  0.4963 -0.1617
#>  -0.0204 -0.2962 -0.4283  0.5512  0.3839  0.1542  0.1853 -0.1114 -0.1726
#>   0.7114 -0.5257 -0.5691  0.4798 -0.1906  0.2108  0.0565 -0.5254 -0.2606
#> 
#> Columns 10 to 18 -0.3426  0.1521  0.6080  0.2891 -0.2006 -0.1015 -0.6547  0.7095 -0.2031
#>   0.0010 -0.3287  0.3676 -0.1687 -0.0967 -0.4770 -0.4011  0.0169 -0.6078
#>  -0.2887 -0.2034 -0.4151  0.1936 -0.1250 -0.4523 -0.6213 -0.2194  0.4377
#> 
#> Columns 19 to 20 -0.5490 -0.1400
#>  -0.1883  0.0024
#>   0.0228 -0.3902
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4946  0.4707  0.2327  0.2931 -0.3486  0.2573  0.2574  0.6971 -0.1110
#>   0.4681  0.2777  0.3355  0.1704  0.0008  0.3576  0.1676  0.5053 -0.1168
#>   0.2939  0.3154  0.4486  0.3067 -0.0420  0.4443  0.4074  0.3560 -0.0950
#> 
#> Columns 10 to 18 -0.6061 -0.3754  0.0054  0.1500  0.2937 -0.2683  0.3698  0.0380 -0.3692
#>  -0.3798 -0.3725 -0.4083  0.2269  0.2623 -0.2974 -0.0191 -0.5680 -0.1990
#>  -0.3157 -0.0189 -0.4742 -0.3299  0.1420 -0.2435 -0.0617 -0.2264  0.1495
#> 
#> Columns 19 to 20  0.7276 -0.0975
#>   0.5028  0.0743
#>   0.7958  0.2401
#> [ CPUFloatType{2,3,20} ]
#>