RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.1994 -0.6349  0.5478  0.3722 -0.3056  0.0746 -0.5223 -0.4682  0.9163
#>  -0.8019 -0.7967  0.4509  0.4495 -0.8775 -0.5374  0.0302  0.5566  0.6127
#>  -0.6907 -0.0730 -0.3114  0.6733 -0.4397 -0.4997 -0.1727 -0.0167 -0.1481
#> 
#> Columns 10 to 18 -0.8818 -0.2283 -0.1134  0.0357 -0.6141  0.3612 -0.0856 -0.1771  0.3784
#>   0.3888 -0.7149  0.0164 -0.5142 -0.6489  0.8913 -0.7622  0.2251 -0.0103
#>  -0.3519 -0.0482  0.2288  0.6289  0.1145 -0.0205 -0.4628  0.2598  0.1725
#> 
#> Columns 19 to 20  0.7145  0.2779
#>   0.4674  0.0551
#>   0.0197  0.6847
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.5016  0.5714  0.3970  0.3111 -0.5423  0.2434 -0.0356  0.4754  0.2284
#>   0.0302 -0.1655 -0.3015  0.4828 -0.0847  0.5113 -0.3308  0.2456  0.0419
#>  -0.1070 -0.0776 -0.2039  0.0901 -0.2827  0.2810  0.2323  0.4898 -0.1335
#> 
#> Columns 10 to 18 -0.0475 -0.2618 -0.5545 -0.1346  0.2891 -0.4909 -0.0620  0.3093  0.6530
#>  -0.7075 -0.6523 -0.0201 -0.0469  0.3551 -0.2050 -0.2243  0.2656  0.2550
#>  -0.5280 -0.5322  0.1174 -0.0424  0.2032  0.0599  0.1284  0.3674  0.2796
#> 
#> Columns 19 to 20 -0.2895  0.5977
#>  -0.0354  0.4285
#>  -0.5405 -0.0438
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3950  0.7463 -0.6029  0.3352  0.0297  0.2381  0.7961  0.3656 -0.4128
#>   0.1810 -0.1639 -0.4615 -0.1812  0.0166  0.5279  0.2833  0.2013 -0.0438
#>   0.5215 -0.1320 -0.0084  0.2810 -0.3805  0.1756 -0.1758  0.0091 -0.0202
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5605  0.1587 -0.1021 -0.2089  0.4377  0.0415 -0.4828  0.3459  0.1075
#>  -0.3420 -0.4318 -0.5234 -0.5357  0.1775  0.7948  0.1686 -0.4741  0.7674
#>   0.3311 -0.5802  0.4235 -0.0789  0.0369 -0.4011  0.1992 -0.3720 -0.3505
#> 
#> Columns 10 to 18  0.1648  0.0420 -0.5305  0.0603  0.4732  0.3053  0.0750  0.1299 -0.1021
#>   0.4326  0.4879  0.0618  0.4537  0.7088  0.7057  0.2332 -0.5225 -0.8174
#>  -0.1445  0.1820  0.0382 -0.0373  0.4555  0.4758  0.0429 -0.5239 -0.0388
#> 
#> Columns 19 to 20  0.0259 -0.1635
#>   0.3935 -0.2209
#>  -0.7000  0.2636
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0467  0.1193 -0.2630  0.4610 -0.3016 -0.0105  0.4041  0.2787  0.0856
#>   0.2563  0.4603 -0.5981  0.2051 -0.1776  0.4959  0.2962  0.0753  0.0220
#>   0.2955  0.1215  0.0461 -0.1131 -0.4338  0.0380  0.0947 -0.0245  0.0910
#> 
#> Columns 10 to 18 -0.0669 -0.2177  0.4449 -0.3688  0.0657 -0.3612  0.1054  0.1810  0.1790
#>   0.1449 -0.2593 -0.3013 -0.4733 -0.3052 -0.0194  0.5961 -0.0889  0.4696
#>   0.0018 -0.0334 -0.1453  0.0734 -0.4981 -0.0972  0.3927 -0.0639  0.1161
#> 
#> Columns 19 to 20  0.0210 -0.2048
#>   0.1801  0.1716
#>   0.0598  0.3980
#> [ CPUFloatType{2,3,20} ]
#>