RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5534  0.2739  0.8061  0.7113  0.0610 -0.6309  0.2649 -0.4260 -0.8475
#>   0.3217  0.0726 -0.1186 -0.7979  0.8222 -0.0809 -0.0620 -0.7560  0.2236
#>   0.7165  0.8268  0.7228  0.3680  0.9113 -0.7138 -0.4047 -0.7526  0.5294
#> 
#> Columns 10 to 18 -0.7142 -0.0014  0.8015  0.6046 -0.2024 -0.8380 -0.4257  0.5126  0.3832
#>   0.0907  0.4626  0.8921 -0.1631 -0.5072 -0.7308  0.0686 -0.8116 -0.2976
#>   0.0151 -0.0056  0.7001  0.8172  0.1358 -0.2974 -0.6539  0.4117  0.8214
#> 
#> Columns 19 to 20 -0.2266 -0.8542
#>   0.2772  0.0251
#>  -0.8234  0.7693
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5406  0.0346  0.3694 -0.1276  0.2705  0.0084 -0.1592  0.0860  0.1650
#>   0.2133  0.4291  0.1077 -0.6225  0.4015 -0.2227 -0.2546 -0.6952 -0.3850
#>  -0.1320 -0.1496  0.0126 -0.0945 -0.4025 -0.3636 -0.3278  0.4039  0.5492
#> 
#> Columns 10 to 18  0.2298  0.2633  0.8239  0.0867  0.2610 -0.8174  0.6806 -0.5123 -0.2308
#>  -0.1679 -0.4281  0.6752  0.1285  0.6623 -0.1983 -0.1295 -0.2466  0.2143
#>  -0.1730 -0.0903  0.6741  0.4192  0.4423 -0.2111 -0.4860  0.1224 -0.2949
#> 
#> Columns 19 to 20  0.5121  0.4671
#>  -0.3144 -0.0945
#>   0.4382  0.6237
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.1109  0.3531  0.4031 -0.3836  0.0111  0.1129 -0.0290 -0.1070 -0.1630
#>   0.0230  0.2152 -0.1435  0.1310  0.6723 -0.2791 -0.4268  0.0442 -0.4111
#>   0.1918  0.2987  0.2954  0.0131 -0.0341  0.0505  0.2667 -0.5130  0.2139
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0187  0.2538  0.6111  0.5033 -0.9017 -0.0278  0.5290  0.6926  0.3371
#>  -0.5481 -0.3449  0.1201  0.0951 -0.6440  0.3515 -0.1459 -0.5893  0.3650
#>  -0.6251 -0.4422 -0.6015  0.4164 -0.1200  0.5953  0.1216 -0.6174  0.2256
#> 
#> Columns 10 to 18  0.4130 -0.0985 -0.0248  0.3803 -0.6128 -0.5376  0.1656 -0.0467  0.5194
#>   0.0067  0.0504 -0.3350 -0.2570 -0.5761 -0.8837  0.3696  0.0706  0.2785
#>  -0.6218 -0.3076  0.1452 -0.2849  0.4333 -0.3424  0.3957  0.4463 -0.7020
#> 
#> Columns 19 to 20 -0.3256 -0.4609
#>  -0.7742 -0.1926
#>  -0.5812  0.4567
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1317  0.2775  0.4267 -0.4361 -0.4102 -0.1069  0.0766 -0.0368 -0.1151
#>  -0.2569  0.3867  0.0791 -0.3006  0.1250  0.4618 -0.3918 -0.3318  0.1692
#>  -0.1825  0.2269  0.1268  0.0859  0.3020  0.0823 -0.2560 -0.4926 -0.0457
#> 
#> Columns 10 to 18 -0.5360 -0.0066  0.3180  0.3254  0.4260  0.0288  0.0683 -0.2844 -0.1952
#>   0.2016  0.1069  0.6232 -0.0526  0.1891 -0.6345  0.7184 -0.3223 -0.2811
#>  -0.3004  0.4245  0.7920  0.2753  0.3002 -0.7067  0.5127  0.2454 -0.2868
#> 
#> Columns 19 to 20  0.5053 -0.4116
#>   0.5713 -0.4549
#>   0.0078 -0.4677
#> [ CPUFloatType{2,3,20} ]
#>