RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0394  0.8515 -0.0109  0.6053  0.1161  0.0708  0.3821 -0.2385  0.1958
#>  -0.7578  0.5795 -0.8703 -0.3670 -0.5331  0.5605 -0.3155 -0.6242 -0.3546
#>   0.7301 -0.7883 -0.4664 -0.8037  0.6437 -0.2718 -0.8692 -0.2892 -0.2131
#> 
#> Columns 10 to 18  0.5245 -0.7259  0.5998 -0.2448  0.3297  0.8271  0.5137 -0.4059  0.4457
#>  -0.8261  0.9088 -0.1651 -0.8077 -0.7212  0.3055  0.9822  0.0850  0.0610
#>  -0.3365  0.3021  0.6001  0.4331  0.3537  0.2257  0.3371  0.3753  0.3158
#> 
#> Columns 19 to 20 -0.5891 -0.8455
#>   0.4149  0.9499
#>   0.7084  0.2649
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0933  0.3907 -0.6706  0.1110  0.0970  0.1531  0.5170 -0.2163 -0.0951
#>   0.3145  0.0087 -0.3005 -0.4049  0.4321  0.0038 -0.2909 -0.2456  0.3743
#>  -0.2386  0.2128  0.6140  0.1400  0.0452 -0.1579 -0.5708  0.3613 -0.3877
#> 
#> Columns 10 to 18 -0.5675  0.1612 -0.2984 -0.7987 -0.5419 -0.4649 -0.4997  0.0822 -0.4354
#>   0.1348 -0.6547  0.1202 -0.3746  0.2797  0.2970 -0.3678  0.0915  0.3989
#>   0.3793 -0.2324  0.6953 -0.0153  0.0639  0.4621  0.0205 -0.0192 -0.7007
#> 
#> Columns 19 to 20 -0.3501 -0.3096
#>  -0.5447 -0.4549
#>   0.2838 -0.5248
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.3543 -0.2345 -0.6994 -0.6262 -0.3045 -0.6368 -0.4031 -0.5040  0.0977
#>  -0.5928  0.2643 -0.2796 -0.0223 -0.3334 -0.6680 -0.3992 -0.2582 -0.3059
#>  -0.5755 -0.2054 -0.1795  0.5502 -0.2716  0.0055 -0.3973 -0.1648 -0.2442
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6159 -0.7063  0.0875 -0.2540 -0.5037 -0.4912 -0.2453  0.4358  0.4591
#>  -0.0010 -0.2611 -0.1841  0.0595  0.0635 -0.3631 -0.0779  0.3634 -0.5514
#>   0.4281 -0.6901 -0.1447  0.4234  0.2050 -0.4514 -0.1800  0.3827 -0.1933
#> 
#> Columns 10 to 18 -0.1230 -0.4245  0.0195 -0.0674 -0.7299  0.5256 -0.3970  0.7333 -0.0856
#>   0.0350  0.0912 -0.2640  0.2707 -0.1314  0.3539 -0.0186 -0.4237  0.0622
#>  -0.2427  0.2368 -0.2435 -0.0981 -0.5399 -0.3089 -0.1277  0.5725  0.1508
#> 
#> Columns 19 to 20  0.1389 -0.4968
#>   0.5125 -0.2802
#>   0.5869 -0.4604
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.5277 -0.5904 -0.2566 -0.1983 -0.3738 -0.4057 -0.1194 -0.3383  0.0541
#>  -0.2824 -0.0971 -0.3384 -0.4098  0.0272 -0.3630 -0.0675 -0.3405 -0.2079
#>  -0.2166 -0.4883 -0.1295 -0.2360  0.1060 -0.0265 -0.1830 -0.3490 -0.2148
#> 
#> Columns 10 to 18 -0.0209  0.1542 -0.0465 -0.3376 -0.2650  0.0028 -0.3069  0.4661  0.5246
#>  -0.4668 -0.0855 -0.1104 -0.2440 -0.0672  0.0036 -0.0248  0.1448  0.2087
#>   0.2122 -0.0261  0.1319 -0.0958 -0.0708  0.4188 -0.2813  0.2923  0.6640
#> 
#> Columns 19 to 20 -0.2399  0.3348
#>  -0.4523  0.1235
#>  -0.4052  0.2416
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>