RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1781 -0.3230 -0.3197  0.5058 -0.4141 -0.7895  0.3329  0.7714  0.3655
#>   0.9166 -0.6370 -0.7058  0.1798 -0.8524  0.5656  0.1922  0.7242 -0.0645
#>   0.2469 -0.3954 -0.8076  0.6952 -0.4730  0.0916  0.8454 -0.6591  0.2451
#> 
#> Columns 10 to 18 -0.4627  0.4086 -0.1373 -0.0813 -0.2223  0.6587  0.6329  0.3083  0.3653
#>  -0.6209 -0.5261  0.0219 -0.9188  0.1328 -0.3269 -0.7358  0.3197 -0.0969
#>  -0.1602  0.6262  0.6118  0.7435 -0.1610  0.8247  0.1391 -0.6065 -0.8544
#> 
#> Columns 19 to 20  0.1187  0.2863
#>  -0.8996 -0.0392
#>  -0.6611  0.3732
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1337  0.5897 -0.2585 -0.5316 -0.1423  0.0168 -0.3217 -0.7133  0.5928
#>  -0.0101 -0.4048 -0.5936 -0.0990 -0.0617  0.3902 -0.3036  0.1597  0.3655
#>   0.1888 -0.7297 -0.4639  0.1754 -0.6526  0.7218 -0.5585 -0.1361 -0.1822
#> 
#> Columns 10 to 18 -0.1789 -0.3862 -0.5578 -0.0943 -0.2186  0.4203  0.1678 -0.4513 -0.2655
#>   0.1979 -0.5307 -0.2340  0.5993 -0.1734  0.4734  0.5909  0.4547 -0.2658
#>  -0.2061 -0.8284  0.1428 -0.0892  0.3802  0.0665 -0.3451 -0.5837 -0.0833
#> 
#> Columns 19 to 20 -0.0412 -0.5729
#>   0.0530  0.0290
#>  -0.7687  0.3883
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0599 -0.1690 -0.1026 -0.1418 -0.1681 -0.0999  0.0854  0.1289  0.3158
#>  -0.3231 -0.4599  0.1303 -0.2672 -0.2475  0.1577 -0.2554  0.0045 -0.3172
#>  -0.2514 -0.7347 -0.3220  0.3693 -0.6129  0.2488  0.1023  0.4195  0.0324
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3899  0.6254 -0.4851  0.3062 -0.0105  0.6980  0.5190  0.0816 -0.2187
#>   0.2350  0.8656 -0.0572 -0.6875 -0.8736  0.7283 -0.0804  0.1075  0.4518
#>   0.4111 -0.4850 -0.7405 -0.1208 -0.2944  0.7429  0.0818 -0.6989  0.0407
#> 
#> Columns 10 to 18  0.2916 -0.2632 -0.5923 -0.2556  0.1514  0.2898 -0.2899 -0.2336 -0.5012
#>  -0.3063 -0.5314 -0.6645  0.4299 -0.8976  0.2740 -0.3954  0.0780 -0.6345
#>   0.3868  0.0225 -0.6144  0.5698 -0.6358  0.1486 -0.4669 -0.2740  0.0202
#> 
#> Columns 19 to 20 -0.4564 -0.1487
#>   0.1967  0.2763
#>  -0.5191 -0.8444
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0692 -0.7214 -0.3522 -0.4825 -0.3721 -0.0242  0.0173  0.3181 -0.0290
#>  -0.2392 -0.4722 -0.7110  0.0843 -0.7815 -0.6562  0.1115  0.5188 -0.0039
#>  -0.0598 -0.0933 -0.0171 -0.5696 -0.2803 -0.2292 -0.1212 -0.0262  0.2756
#> 
#> Columns 10 to 18  0.1306 -0.5852 -0.3118  0.0848 -0.0167  0.1447 -0.3744 -0.1723 -0.3969
#>   0.1900 -0.4948  0.1681  0.0788 -0.1415  0.0124 -0.4193  0.1631 -0.2778
#>   0.3155 -0.6526 -0.5125  0.4265 -0.4000 -0.1723 -0.4085 -0.2263 -0.2964
#> 
#> Columns 19 to 20 -0.0261  0.4431
#>  -0.0611  0.2504
#>   0.0529  0.3394
#> [ CPUFloatType{2,3,20} ]
#>