RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.8107 -0.6149 -0.5709 -0.5249 -0.6257  0.7747 -0.7744  0.2835  0.6211
#>  -0.0455 -0.6434  0.6417  0.2375 -0.3229  0.1271 -0.2809  0.2194 -0.2369
#>   0.2656 -0.4615 -0.0645 -0.5447 -0.2307 -0.5694 -0.1902  0.7385  0.4502
#> 
#> Columns 10 to 18 -0.8309 -0.7710 -0.9425  0.6024 -0.3747 -0.3119 -0.3431  0.1497  0.7140
#>  -0.3902  0.5517  0.3542  0.1979  0.0447  0.5134  0.4999  0.0956  0.1276
#>   0.2519  0.6524  0.4916  0.8927 -0.1563  0.0457  0.5219 -0.0006 -0.1248
#> 
#> Columns 19 to 20  0.6515  0.8888
#>   0.2828  0.3045
#>   0.0437  0.4198
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.2652 -0.1238  0.4274  0.8308  0.4785  0.2981  0.5312  0.4404 -0.0657
#>  -0.1815 -0.5366  0.0249  0.4434 -0.0695  0.2182 -0.0715 -0.1878 -0.2879
#>  -0.4402 -0.6833  0.1033  0.3330 -0.1479  0.6464  0.1481  0.3734 -0.4610
#> 
#> Columns 10 to 18  0.0912  0.1656  0.4578 -0.3799  0.3233 -0.4235 -0.0463  0.2725 -0.4699
#>   0.1087 -0.2229 -0.5180 -0.1811  0.3843  0.3148 -0.0672  0.0469  0.2432
#>  -0.0161 -0.0922 -0.3610 -0.0830  0.4566  0.5559  0.1271  0.3140  0.3320
#> 
#> Columns 19 to 20 -0.1421 -0.4125
#>   0.4314 -0.3121
#>  -0.2959  0.2856
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.1638 -0.7302 -0.1974  0.0720 -0.5511 -0.3550 -0.2559  0.6448 -0.4292
#>  -0.1916 -0.4872 -0.0719  0.1503 -0.2948 -0.3889 -0.0032  0.3105 -0.5216
#>  -0.0428 -0.6017 -0.2370  0.0032  0.2865  0.1294 -0.3251 -0.1192 -0.3854
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.6267  0.3895  0.3366  0.5321  0.3006 -0.1499 -0.3479 -0.0806 -0.0525
#>   0.7420 -0.0659 -0.2984  0.5814  0.0737 -0.1258 -0.1870  0.4126 -0.4902
#>   0.6720 -0.2076 -0.3789 -0.1303 -0.2764  0.3714  0.0897  0.1772 -0.1589
#> 
#> Columns 10 to 18 -0.2297  0.0262  0.1405  0.6856  0.0130 -0.1358 -0.0128  0.5421  0.3448
#>   0.3289 -0.1225  0.2415  0.4729  0.4652  0.0806  0.1034  0.4153  0.3813
#>   0.4152 -0.1237 -0.1496 -0.3860  0.6823  0.2739  0.2131 -0.1758 -0.0044
#> 
#> Columns 19 to 20 -0.4008  0.7795
#>  -0.0292  0.7869
#>   0.0222  0.3859
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0526  0.2448 -0.1835  0.2066 -0.3216  0.0260 -0.0647 -0.1191 -0.3555
#>  -0.2371 -0.0354 -0.0537 -0.0576  0.0056  0.1063 -0.0450 -0.0282 -0.5574
#>  -0.3150 -0.5654 -0.3008  0.3325 -0.0564  0.2021  0.0212  0.3454 -0.3995
#> 
#> Columns 10 to 18  0.5548 -0.1898 -0.1724 -0.2181  0.2796 -0.1159 -0.3300 -0.4305  0.1675
#>   0.5388 -0.1753 -0.1029  0.0171  0.1962  0.1596 -0.2601 -0.1905  0.1161
#>   0.1854 -0.1672 -0.3049  0.1239  0.2391  0.1062  0.2020 -0.2652 -0.0940
#> 
#> Columns 19 to 20 -0.4807 -0.3330
#>  -0.2323 -0.2194
#>  -0.3605 -0.0839
#> [ CPUFloatType{2,3,20} ]
#>