RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3188  0.0571  0.7245  0.4925  0.3861 -0.0432  0.1556 -0.1220 -0.8936
#>  -0.6829  0.3030  0.2583 -0.0666 -0.3218 -0.1492  0.5181 -0.9709  0.4974
#>   0.2871 -0.6584  0.6260 -0.3944 -0.4255  0.3791  0.4837 -0.6951 -0.2234
#> 
#> Columns 10 to 18 -0.8514 -0.0769 -0.6643  0.0144  0.3087 -0.8586 -0.5629  0.2088  0.8832
#>   0.8487 -0.3397  0.0350  0.3448 -0.3914  0.4118 -0.2779  0.3363  0.5774
#>   0.0899 -0.4279  0.5160  0.4551 -0.5288 -0.1956  0.0096 -0.1626  0.6064
#> 
#> Columns 19 to 20 -0.5613  0.5193
#>   0.8036  0.2785
#>   0.2026  0.6291
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.7395  0.7358  0.3672  0.3415 -0.1636 -0.1214 -0.5354  0.0922 -0.4402
#>   0.3924  0.7281  0.4001  0.2411  0.5274  0.3488 -0.2558  0.5592 -0.2928
#>   0.3194  0.3605  0.4595 -0.3551  0.6415 -0.0735  0.4840  0.5758 -0.3400
#> 
#> Columns 10 to 18 -0.1625 -0.3934  0.2853  0.1139 -0.6813 -0.5333 -0.2426  0.6121  0.5008
#>  -0.0634 -0.3891 -0.4198 -0.3757 -0.0167  0.4014 -0.0981  0.0518  0.1498
#>  -0.0831 -0.4525 -0.4178 -0.3842 -0.1647  0.0625 -0.0041 -0.3560  0.7586
#> 
#> Columns 19 to 20 -0.0918 -0.1192
#>  -0.3000 -0.0380
#>  -0.2074  0.3965
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0043  0.7657  0.1102 -0.1281 -0.3425 -0.2985  0.1473  0.1693 -0.5075
#>  -0.4604  0.8360  0.4947  0.3127 -0.1296 -0.1938  0.1077 -0.1862 -0.2750
#>   0.8191  0.6962  0.0783  0.5523 -0.3353  0.1738 -0.5928  0.1584 -0.5501
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.3703  0.0288  0.6673  0.2027 -0.1045 -0.3488 -0.6151 -0.2207 -0.2066
#>  -0.2235  0.7016  0.9348  0.7642 -0.8184  0.0126  0.3995  0.5572 -0.5477
#>  -0.1038 -0.6468  0.7574 -0.7476 -0.1708  0.3770  0.4383 -0.7234 -0.0168
#> 
#> Columns 10 to 18  0.5328 -0.2358  0.0635 -0.1291  0.7060  0.0966 -0.5983  0.4711  0.2416
#>   0.3413 -0.2648 -0.5765  0.6415  0.0204  0.8095  0.1340 -0.6246  0.8473
#>   0.1423 -0.8081  0.5746  0.0820 -0.3834 -0.1588  0.0165  0.1149  0.5990
#> 
#> Columns 19 to 20 -0.1704  0.3200
#>   0.8067 -0.5908
#>  -0.4756  0.0133
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.4419  0.7848  0.3040  0.2226 -0.1644  0.0719 -0.3523 -0.0063 -0.4541
#>  -0.1203  0.6206  0.3378 -0.1908  0.4575 -0.0020  0.0005 -0.3075 -0.4125
#>   0.7667  0.6701 -0.1683  0.2543 -0.3683  0.2026 -0.5519 -0.5009 -0.1387
#> 
#> Columns 10 to 18  0.2709 -0.3103 -0.2045 -0.3960  0.0076  0.4404  0.1525  0.1205  0.0125
#>   0.0452 -0.6913 -0.6401 -0.4049 -0.2271  0.1431 -0.3131 -0.3269  0.2184
#>   0.2114 -0.5768 -0.2395 -0.4350 -0.5052  0.2361 -0.4345  0.4174  0.1239
#> 
#> Columns 19 to 20  0.0137  0.2350
#>   0.0361  0.2089
#>   0.0179  0.4053
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>