RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size
The number of expected features in the input x

hidden_size
The number of features in the hidden state h

num_layers
Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity
The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias
If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first
If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout
If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional
If TRUE, becomes a bidirectional RNN. Default: FALSE

...
other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)

}

#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0431 -0.2139 -0.2361 -0.4504  0.5202  0.1769 -0.3639 -0.3190 -0.0715
#>   0.3984 -0.2554 -0.9418  0.1586  0.8787  0.5664  0.3721 -0.5942 -0.6603
#>   0.1056 -0.1448 -0.0515 -0.1270 -0.8795 -0.2447 -0.8512  0.2739  0.3198
#> 
#> Columns 10 to 18 -0.4905 -0.8149  0.3254 -0.4250 -0.3702  0.7061 -0.3324  0.7302  0.3537
#>  -0.5270 -0.6441  0.1013 -0.6162  0.1196  0.1362  0.8695  0.3156  0.7666
#>  -0.5993 -0.9111  0.5033 -0.1506  0.1323  0.7086 -0.4128  0.8824 -0.6607
#> 
#> Columns 19 to 20 -0.5831 -0.1006
#>   0.4028 -0.4089
#>   0.7756 -0.4328
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.2488 -0.6372 -0.1901 -0.1550  0.1152 -0.5678 -0.7414  0.0012 -0.0759
#>   0.2733 -0.4912 -0.5381 -0.2039  0.5311 -0.5709 -0.3336  0.1338 -0.3571
#>  -0.4251 -0.2742  0.1339  0.0268 -0.3235 -0.5340 -0.5196  0.1567 -0.2058
#> 
#> Columns 10 to 18 -0.1255 -0.0384  0.1538 -0.6606 -0.2143  0.3466  0.0878  0.6044  0.6205
#>   0.2153  0.1945 -0.0084 -0.3608 -0.3178  0.0971  0.4773  0.2741  0.4125
#>   0.0257 -0.4267  0.4899 -0.3417 -0.4260 -0.0716  0.6126  0.2800 -0.2656
#> 
#> Columns 19 to 20  0.2370 -0.2524
#>   0.4538 -0.2181
#>   0.3769  0.1823
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.5813 -0.4664  0.0495  0.0822  0.1386 -0.4266 -0.6595 -0.0789 -0.2586
#>  -0.3802 -0.0114 -0.2891  0.1410  0.5044  0.2871 -0.6034 -0.2390 -0.0389
#>  -0.4193  0.1499 -0.2558  0.0988 -0.1229 -0.2918 -0.6023  0.1550 -0.3054
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0947  0.3838  0.1330 -0.7024 -0.1966  0.6087 -0.3917 -0.5802  0.4728
#>  -0.4994 -0.1207 -0.3186 -0.7583 -0.1387  0.2366  0.0349 -0.6574  0.2730
#>  -0.0540  0.3153 -0.5581  0.0707  0.1786  0.1226  0.5738 -0.2252  0.6478
#> 
#> Columns 10 to 18  0.2966 -0.3210  0.2356 -0.7075  0.5515 -0.0347  0.2067 -0.6198 -0.0918
#>   0.3902 -0.3271  0.5806 -0.6917  0.5421  0.4157  0.1605 -0.5318 -0.4165
#>  -0.5083  0.2999  0.3961  0.0883  0.5380 -0.5236  0.2739  0.4166 -0.2360
#> 
#> Columns 19 to 20  0.2070 -0.5402
#>   0.0060  0.1133
#>   0.0447  0.4647
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4069 -0.5259 -0.0513  0.2366  0.4100 -0.2709 -0.1122 -0.0171 -0.2253
#>  -0.3465 -0.5106 -0.0360  0.2461  0.4105  0.0040 -0.2163  0.0062 -0.1924
#>  -0.0051  0.3442 -0.3320  0.2566 -0.0859 -0.0118 -0.3621  0.1946 -0.0884
#> 
#> Columns 10 to 18 -0.1671 -0.2384 -0.1285 -0.0218 -0.0159 -0.5240  0.4843  0.0305 -0.1829
#>   0.2003  0.0230 -0.4368 -0.1624 -0.2206 -0.4454  0.5737  0.1230  0.0482
#>  -0.2729 -0.2799  0.0000 -0.3067  0.2167 -0.2683 -0.0297  0.4752  0.3098
#> 
#> Columns 19 to 20 -0.0240  0.2732
#>  -0.0184  0.0321
#>   0.1995  0.1558
#> [ CPUFloatType{2,3,20} ]
#>