RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.5079  0.0171 -0.2103  0.8954  0.4582 -0.6840 -0.7239 -0.0367 -0.8811
#>   0.9094  0.7721 -0.7149  0.6441  0.1174 -0.4713  0.1105 -0.2006 -0.1700
#>   0.7066  0.2627 -0.7612 -0.5210  0.6456  0.5314 -0.1989 -0.3660  0.1459
#> 
#> Columns 10 to 18 -0.2786  0.4058  0.2725 -0.3738 -0.1395 -0.2796 -0.4478 -0.9197 -0.6304
#>  -0.4226  0.2370  0.3522  0.4004 -0.3815  0.5640  0.6602  0.6470 -0.2487
#>  -0.0082  0.0097 -0.0865 -0.2609  0.2133 -0.1909  0.8110 -0.1423  0.4260
#> 
#> Columns 19 to 20 -0.5898  0.2297
#>  -0.0190  0.3477
#>   0.2264 -0.3893
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6729 -0.4967 -0.4755  0.5084  0.0553 -0.1420 -0.4336 -0.7860 -0.1329
#>  -0.6203 -0.5209  0.0538  0.0161 -0.3072  0.4530 -0.5028  0.0997 -0.3890
#>   0.0189 -0.4507 -0.1519  0.4843 -0.2699  0.2314 -0.1959 -0.1865  0.1539
#> 
#> Columns 10 to 18 -0.2887 -0.6446  0.6415 -0.1562  0.3315 -0.0838  0.3578 -0.3374 -0.1457
#>   0.6490 -0.2298  0.8080 -0.2510  0.3190  0.3048  0.4243  0.4003 -0.3844
#>   0.2181 -0.7087  0.8901  0.0927 -0.0590 -0.0851  0.4548  0.1939 -0.3610
#> 
#> Columns 19 to 20  0.1994 -0.4917
#>  -0.3533 -0.3213
#>  -0.4026 -0.1330
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.0360 -0.7932  0.1956  0.2193 -0.4253 -0.0153 -0.1011 -0.4853  0.2906
#>  -0.0057  0.0539 -0.2461  0.4007  0.3487  0.3527  0.0655 -0.1778  0.0073
#>   0.1555 -0.5438  0.1309  0.0601 -0.4539  0.2940  0.1523  0.1321 -0.0275
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.9088  0.1990 -0.1666 -0.9178 -0.2628  0.1908  0.5707 -0.0566 -0.6487
#>  -0.1965  0.7025  0.2165 -0.2818  0.7598 -0.1314  0.3048 -0.3956  0.2206
#>  -0.8503  0.3169  0.3926 -0.3742  0.4503 -0.4748 -0.1085  0.0433 -0.2370
#> 
#> Columns 10 to 18 -0.2442 -0.0892 -0.4129 -0.2057  0.2715  0.3739 -0.6366  0.4411 -0.6213
#>   0.0920 -0.1616 -0.0070  0.4422 -0.0572  0.2779 -0.1588 -0.0905  0.1902
#>  -0.6907 -0.1666  0.4438  0.4155  0.0414  0.0792  0.0264  0.0977 -0.0537
#> 
#> Columns 19 to 20 -0.1976  0.0294
#>   0.4701 -0.5283
#>   0.4742  0.2853
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0707 -0.6273  0.3114  0.1182 -0.6497  0.2477  0.2238  0.0988  0.3522
#>  -0.0914 -0.3376 -0.2153  0.3762 -0.1779 -0.4930 -0.2267 -0.0100 -0.2371
#>  -0.0199  0.0783 -0.1059  0.1960 -0.1585 -0.0284 -0.0848  0.2777  0.0664
#> 
#> Columns 10 to 18  0.3553 -0.5757  0.6985 -0.1915  0.2785 -0.3568  0.2894  0.4946  0.0212
#>   0.3885  0.1943  0.3645 -0.1446 -0.5064  0.0555  0.3696  0.2361 -0.5584
#>   0.1536  0.1259  0.2637 -0.1221  0.0809 -0.1124  0.2286  0.3023 -0.2118
#> 
#> Columns 19 to 20  0.1852  0.0700
#>   0.1496  0.2424
#>   0.1645  0.4499
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>