RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.5667 -0.1397  0.7498  0.7778  0.3602 -0.5265 -0.4283 -0.5213 -0.8292
#>  -0.1084 -0.0464  0.5597  0.0621  0.1367  0.3887  0.2729  0.1935  0.0687
#>  -0.5698  0.5830 -0.1892  0.6360  0.5580  0.7901  0.6185 -0.1671  0.2899
#> 
#> Columns 10 to 18 -0.2446 -0.2902  0.1283  0.3794  0.6443 -0.5181 -0.8813 -0.2083  0.1131
#>  -0.4994 -0.7451  0.5305 -0.3218  0.3208 -0.3731  0.4423  0.0567  0.1874
#>   0.0443  0.2495  0.0112 -0.4349  0.5507  0.5159  0.5024 -0.5309 -0.3965
#> 
#> Columns 19 to 20  0.1220  0.5621
#>   0.7808  0.2143
#>  -0.0004  0.3319
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1220  0.6054 -0.2041  0.5396 -0.6470  0.2787  0.1143 -0.7746  0.2164
#>   0.2306 -0.2501  0.1762  0.5530 -0.5366  0.5464 -0.3246 -0.2805 -0.2066
#>   0.1194  0.3922  0.0556  0.3182  0.4173  0.6278  0.5831 -0.1837 -0.3460
#> 
#> Columns 10 to 18 -0.1905 -0.3353 -0.1078  0.6206 -0.3964  0.5824 -0.1469 -0.4854  0.6087
#>   0.4424 -0.5179  0.4636 -0.2376  0.1072 -0.0950  0.2593  0.2416  0.3412
#>   0.4361  0.0517  0.3439 -0.2024  0.0669  0.6988 -0.3355  0.4153  0.6230
#> 
#> Columns 19 to 20  0.2550 -0.1418
#>   0.3255  0.1107
#>  -0.2223  0.6052
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.4272  0.6646  0.0953  0.7133 -0.6495  0.4329  0.2233 -0.5005 -0.2471
#>  -0.2850  0.0611  0.4207  0.0854 -0.2010  0.5749  0.2485  0.1846 -0.0744
#>  -0.2354  0.5553  0.4162  0.2332 -0.1454  0.3775  0.4112  0.0722 -0.2030
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3382 -0.8774 -0.1105 -0.3874  0.4653  0.5083 -0.2305  0.4001  0.0264
#>  -0.3747 -0.1310 -0.0826 -0.0629  0.8338  0.4534  0.4580 -0.0979 -0.4581
#>   0.5880  0.4527  0.8013  0.1221  0.0236  0.7163  0.3368  0.2022 -0.4137
#> 
#> Columns 10 to 18  0.6812 -0.6197 -0.2749 -0.7083 -0.3284  0.0623 -0.1004 -0.0898  0.8060
#>   0.8707  0.0185  0.3765 -0.4500 -0.0498  0.1252 -0.2945  0.5802  0.3039
#>   0.2845  0.6555 -0.6110  0.4946 -0.4785  0.1866 -0.1915  0.5090  0.1768
#> 
#> Columns 19 to 20  0.4481 -0.6187
#>   0.7527 -0.2606
#>  -0.2026 -0.6157
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0002 -0.2309  0.0518  0.2922 -0.2014  0.5587  0.1827 -0.0137 -0.1575
#>  -0.3220  0.2949  0.1957  0.3019 -0.6725  0.5660  0.2538 -0.3589 -0.1053
#>  -0.0499  0.5084  0.7128  0.6341  0.2169  0.5055  0.1228  0.3597 -0.4060
#> 
#> Columns 10 to 18  0.1317 -0.0503 -0.5563  0.1519  0.2237  0.2898 -0.4007  0.5627  0.6534
#>  -0.1715  0.1302 -0.0409  0.2044  0.4189  0.2663 -0.6099  0.3867  0.5151
#>  -0.1591 -0.0169  0.1685 -0.1935  0.4485  0.6799 -0.2922  0.7184  0.5883
#> 
#> Columns 19 to 20  0.3871  0.5856
#>   0.4643  0.5173
#>   0.4152  0.2383
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>