RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.3975 -0.7605 -0.5376  0.5728  0.6099  0.0247  0.5401  0.6284  0.5689
#>  -0.6417  0.5432  0.4602 -0.6703  0.1084  0.2441 -0.2016  0.6287  0.4088
#>   0.5463  0.7303 -0.8516 -0.3765  0.8723  0.0128 -0.8741  0.8766  0.0154
#> 
#> Columns 10 to 18  0.1723 -0.1455 -0.8941  0.9117  0.5198  0.4230 -0.5789 -0.2467  0.2713
#>  -0.0307  0.1823  0.1823 -0.0230 -0.8740 -0.7792  0.1342  0.4872  0.3864
#>   0.3627 -0.6493  0.7308  0.1623 -0.0532  0.1234  0.3035  0.6759 -0.9320
#> 
#> Columns 19 to 20 -0.3071  0.1128
#>  -0.8393 -0.2570
#>  -0.2630 -0.5766
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1785  0.1163  0.4587  0.1917  0.3057  0.1130 -0.5265 -0.1521 -0.3261
#>   0.2663  0.2277 -0.2943  0.1427  0.7470 -0.3136 -0.0825  0.2439  0.1611
#>  -0.3262  0.3823  0.3357 -0.3887  0.9263 -0.4839  0.6401 -0.1047  0.4889
#> 
#> Columns 10 to 18  0.1006 -0.3311 -0.2075  0.2146  0.3596  0.5334 -0.3882 -0.3213  0.3511
#>   0.5092 -0.7780  0.1044 -0.2610 -0.6466 -0.3099 -0.4103  0.5948  0.3680
#>   0.5780 -0.8353 -0.5400  0.0029  0.0041 -0.0345 -0.2851 -0.4438 -0.0872
#> 
#> Columns 19 to 20 -0.3809  0.1248
#>  -0.2425 -0.5149
#>  -0.3585 -0.7738
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.4292 -0.1892  0.0264 -0.1522  0.3950 -0.3588 -0.0867  0.1730  0.7450
#>   0.0263 -0.2462 -0.6485 -0.4618  0.6268 -0.4202 -0.0022  0.2054  0.3046
#>  -0.1189  0.0512  0.3139  0.5285  0.7421  0.0674 -0.2593 -0.5488  0.9411
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0355 -0.2383  0.2154  0.6138  0.1016 -0.5188 -0.7216 -0.3858 -0.4688
#>  -0.4168 -0.0040  0.1038 -0.2239  0.5347  0.1059  0.6409  0.0014 -0.0209
#>  -0.4133 -0.4469  0.5111  0.0111  0.1689  0.5007 -0.4659 -0.5111 -0.7667
#> 
#> Columns 10 to 18  0.3593  0.2855  0.1170  0.2670 -0.0398  0.2671  0.3034 -0.2408 -0.4583
#>   0.1749  0.6577 -0.2240  0.5829  0.2487 -0.3972  0.1581 -0.6134  0.1268
#>  -0.0361 -0.1948  0.4413 -0.7486  0.7364  0.5632  0.4294 -0.2594  0.4168
#> 
#> Columns 19 to 20  0.2072  0.7036
#>   0.7489  0.4858
#>   0.6328  0.6655
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.0053  0.2013 -0.0451 -0.0835  0.7270 -0.4472 -0.0556  0.0973  0.7373
#>  -0.5252 -0.1195 -0.5261 -0.1097  0.6265 -0.0920  0.2746 -0.0315  0.8305
#>  -0.2219  0.0783  0.4800 -0.4087  0.5087  0.3140  0.2199 -0.0933  0.4918
#> 
#> Columns 10 to 18  0.1677 -0.4223 -0.3101 -0.5531 -0.2294 -0.4500 -0.1205  0.1034 -0.2615
#>  -0.3063 -0.4224 -0.4217  0.2074 -0.5034  0.3802 -0.0570  0.3239 -0.6001
#>   0.2206 -0.3814 -0.3311 -0.2808 -0.2900 -0.3628  0.3161  0.0442 -0.0789
#> 
#> Columns 19 to 20 -0.2661 -0.3064
#>  -0.0686 -0.6832
#>  -0.5569  0.0085
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>