RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.4249  0.2309  0.6895  0.2988 -0.3328 -0.4446 -0.6144 -0.7676  0.1543
#>  -0.3558 -0.0005  0.6854  0.0895  0.0795  0.1864 -0.7665  0.0994  0.5267
#>   0.8179  0.8219  0.7162  0.9398  0.3669 -0.6735 -0.4383  0.2547  0.8445
#> 
#> Columns 10 to 18 -0.6896  0.8143 -0.1963  0.1189 -0.4419 -0.7277 -0.4469  0.5066  0.2986
#>  -0.7509  0.7152  0.8282 -0.3938 -0.1216 -0.0522  0.5522 -0.4416  0.5408
#>   0.2589 -0.6175  0.2332 -0.5832  0.3402  0.2308 -0.8097  0.5616 -0.7446
#> 
#> Columns 19 to 20  0.5877  0.6177
#>   0.5951  0.4205
#>  -0.7224  0.0182
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.0107  0.5370  0.1376  0.4478  0.5364 -0.2500 -0.3526  0.5784  0.0928
#>   0.6043 -0.2841  0.2471  0.4743  0.4873 -0.1402 -0.7336 -0.2042  0.4918
#>   0.6766 -0.0170  0.1903  0.4713  0.2891 -0.5534 -0.0954  0.3595  0.0658
#> 
#> Columns 10 to 18  0.0660  0.1555  0.2492  0.6953  0.7404 -0.2657 -0.4413  0.1401  0.1694
#>   0.0731  0.6540  0.4241  0.0231  0.5732 -0.3168  0.1323 -0.3921 -0.1037
#>  -0.1170  0.2773 -0.1866  0.5586  0.6137 -0.0352 -0.3438  0.4442 -0.0188
#> 
#> Columns 19 to 20 -0.4113  0.3443
#>   0.1904  0.2092
#>  -0.0630  0.5944
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.2268  0.0644 -0.4543  0.1333  0.3581 -0.0593  0.1868  0.2850 -0.2659
#>  -0.1226 -0.4565  0.0265  0.2058  0.6604 -0.1317 -0.3574 -0.1915 -0.2152
#>   0.4132 -0.0301 -0.3907  0.2237  0.1424 -0.3523  0.1384 -0.1982 -0.0274
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.7627 -0.5833  0.5170 -0.0300 -0.2940  0.0969 -0.1463  0.4346 -0.3626
#>  -0.2002  0.4236 -0.2136  0.1174 -0.0801  0.3151 -0.5025 -0.2996  0.1781
#>  -0.3376  0.3231 -0.1847 -0.3717  0.6292 -0.0677 -0.0739 -0.7511 -0.0285
#> 
#> Columns 10 to 18  0.2522 -0.2687 -0.2230 -0.2873  0.2420 -0.6553  0.3003  0.5461  0.2988
#>  -0.5996  0.0944 -0.4412  0.2179  0.3105  0.3196  0.1272  0.4394 -0.5070
#>   0.3814  0.3919 -0.1359  0.5498  0.7629  0.0574  0.8627 -0.2084  0.7002
#> 
#> Columns 19 to 20  0.8096  0.4065
#>  -0.0956 -0.3458
#>   0.2486 -0.1643
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3610  0.0053 -0.4444  0.3919  0.6828  0.0038  0.1679  0.0409 -0.5842
#>   0.2370 -0.0638 -0.1282  0.4964  0.3289  0.0361  0.0814  0.1360 -0.6515
#>   0.4241  0.1806 -0.3898  0.5932  0.4628 -0.1602  0.3958 -0.1482 -0.6018
#> 
#> Columns 10 to 18  0.1250  0.4874 -0.4335  0.4392 -0.5667 -0.3274 -0.5394 -0.2067 -0.2702
#>   0.0599  0.5484 -0.1376  0.5426  0.0574  0.0059 -0.4771 -0.2143 -0.0536
#>   0.1399  0.3182  0.0085 -0.0605  0.0538  0.0765 -0.5611 -0.3101  0.1029
#> 
#> Columns 19 to 20 -0.1870 -0.1716
#>   0.3609  0.6210
#>  -0.2251  0.2186
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>