RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.2520 -0.0365  0.7817 -0.2513  0.2332 -0.0323  0.2506  0.2973 -0.4732
#>   0.6957  0.4773  0.8181  0.5206 -0.3076 -0.5178  0.4526  0.8690 -0.9342
#>   0.0580 -0.0792  0.6892 -0.0494  0.8078 -0.4430  0.7210  0.7374  0.7074
#> 
#> Columns 10 to 18 -0.3163 -0.8452  0.3288  0.0374 -0.4921  0.1078 -0.2724 -0.8069 -0.9533
#>   0.2143 -0.5784 -0.7898  0.4269  0.7524  0.2665 -0.6304 -0.1848 -0.8226
#>   0.5679 -0.8107 -0.5899  0.6687 -0.4633 -0.6310  0.7944  0.6666  0.6120
#> 
#> Columns 19 to 20  0.6496  0.3634
#>  -0.0799  0.7867
#>   0.4449  0.5647
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.3886 -0.2954  0.1373 -0.0020 -0.0124 -0.6599  0.2316  0.0381 -0.4500
#>   0.5973 -0.0618  0.3490  0.2903 -0.6748 -0.4950 -0.0585  0.3384 -0.2666
#>   0.4040 -0.4654  0.2116 -0.0039  0.5631 -0.4305  0.0150 -0.4690  0.2531
#> 
#> Columns 10 to 18  0.5263 -0.5076 -0.3685 -0.5089  0.6884 -0.1130  0.0163 -0.3123 -0.0628
#>  -0.2278 -0.2949 -0.6205 -0.6921  0.7087  0.0495 -0.3662  0.4350  0.4845
#>   0.4218  0.4831 -0.3230 -0.3740  0.1889 -0.0616  0.0510 -0.6987 -0.2230
#> 
#> Columns 19 to 20  0.2190  0.0709
#>  -0.1026  0.0558
#>  -0.1620 -0.4777
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0755  0.1516 -0.0276  0.3713 -0.6805 -0.3070 -0.1796  0.3856 -0.2968
#>   0.7035 -0.1844  0.3009  0.0205 -0.6655 -0.0067  0.0599  0.1885 -0.0977
#>  -0.6266 -0.0416 -0.1720  0.4161  0.2162 -0.1547 -0.5208 -0.1535 -0.0036
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.0870  0.3590 -0.6513  0.5629  0.4331 -0.3247  0.0181  0.5665 -0.1977
#>   0.5617 -0.0647  0.1467 -0.6981 -0.5163  0.6176 -0.4675  0.7732  0.6002
#>   0.1109 -0.2969  0.4862 -0.3473  0.3223  0.7110 -0.2147 -0.7983  0.7517
#> 
#> Columns 10 to 18 -0.0691  0.2943 -0.5739  0.1370 -0.5316  0.2580  0.1147 -0.1859 -0.1010
#>   0.0625  0.5383  0.1546 -0.6062  0.0163  0.1324  0.5914  0.6763 -0.1369
#>   0.8039  0.6951 -0.3523 -0.6646  0.5595  0.7257  0.7435  0.6418  0.3231
#> 
#> Columns 19 to 20  0.2005 -0.0272
#>  -0.3682  0.4097
#>  -0.4769  0.2303
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.6237 -0.1229  0.6091 -0.0737 -0.3778 -0.2592 -0.0696 -0.1657 -0.1975
#>  -0.1744 -0.1457  0.2208  0.0126  0.0830 -0.1485 -0.3424  0.3647 -0.4849
#>  -0.3726 -0.1521 -0.4557  0.6781 -0.5307 -0.0863 -0.3797  0.1494 -0.0104
#> 
#> Columns 10 to 18 -0.2485 -0.3284 -0.3314 -0.1299  0.2591 -0.1100  0.0461 -0.0743 -0.5194
#>  -0.1333 -0.5630  0.1494 -0.1776  0.1069 -0.4108  0.0636 -0.3428  0.3802
#>   0.5926 -0.2381 -0.6550  0.1670  0.7801 -0.2598  0.1502 -0.2905  0.6811
#> 
#> Columns 19 to 20 -0.2705  0.1660
#>   0.0679  0.1948
#>  -0.0152 -0.0093
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>