RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.1505 -0.5665  0.1897 -0.5215  0.5671  0.0850  0.1419  0.8084  0.3325
#>  -0.4489 -0.3908 -0.5197  0.7891  0.8895 -0.5544  0.2084 -0.7723 -0.0664
#>   0.0973  0.0851 -0.7721  0.1633 -0.0994 -0.6027 -0.9249 -0.1244  0.8644
#> 
#> Columns 10 to 18  0.3545  0.7279  0.2403  0.8744  0.0500  0.5865  0.4844  0.1082 -0.1784
#>   0.9075 -0.0352  0.0880 -0.1323  0.5380 -0.6120  0.0957  0.4594  0.3706
#>   0.3996 -0.7154  0.0490  0.6858 -0.3888  0.3355 -0.1109  0.1685  0.5838
#> 
#> Columns 19 to 20  0.5959 -0.3816
#>   0.3303 -0.3262
#>  -0.3168  0.5532
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.1364  0.0416  0.0481  0.5172  0.1335 -0.0736 -0.2736  0.2653  0.6043
#>  -0.0759  0.3396 -0.0047 -0.3880  0.2909  0.2583 -0.4566  0.1061  0.2905
#>  -0.1463  0.1324 -0.6922 -0.3375 -0.0814  0.2357 -0.3817  0.2521  0.8662
#> 
#> Columns 10 to 18  0.4347  0.2509  0.0156 -0.1406 -0.1988  0.4360  0.2932 -0.0524  0.4331
#>   0.1499 -0.6191 -0.1474  0.5052  0.3061  0.4389  0.1013  0.6084 -0.0633
#>   0.2299 -0.3766  0.1626  0.0624  0.5790  0.6588  0.1280  0.1766  0.3202
#> 
#> Columns 19 to 20  0.3393 -0.1566
#>  -0.4633 -0.7576
#>  -0.4300  0.1713
#> 
#> (3,.,.) = 
#>  Columns 1 to 9  0.0725 -0.0595 -0.4664  0.1125  0.2858  0.2331  0.3603  0.0253  0.5534
#>   0.1854  0.0114  0.0807 -0.0806  0.2056 -0.0879 -0.0650  0.0236  0.4775
#>  -0.0117 -0.4360 -0.3221 -0.2238  0.0748 -0.2202 -0.1251  0.1099  0.7227
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.0257  0.3095  0.4908 -0.4591  0.0343  0.2881 -0.5138  0.2353  0.4179
#>   0.3675 -0.7344  0.2869 -0.8995  0.5388 -0.4722 -0.5391 -0.6268 -0.6098
#>   0.2473 -0.5190  0.0350  0.3299 -0.0405 -0.4302 -0.7747 -0.5839 -0.1352
#> 
#> Columns 10 to 18 -0.6876 -0.1567 -0.7185  0.2936  0.5779 -0.2755 -0.2908 -0.3028 -0.6641
#>  -0.0143 -0.2891  0.1754  0.2697 -0.6515  0.1122  0.0597 -0.2069  0.1951
#>  -0.0761  0.1960 -0.3378 -0.2266 -0.5480  0.4694 -0.4272 -0.1155  0.3320
#> 
#> Columns 19 to 20  0.2165 -0.1638
#>   0.8332  0.0892
#>   0.4105 -0.5244
#> 
#> (2,.,.) = 
#>  Columns 1 to 9  0.1357  0.0563 -0.1381  0.1174  0.1743 -0.2022 -0.0865  0.1568  0.2190
#>  -0.1784 -0.2968 -0.1057 -0.3717  0.3915  0.2854  0.2520  0.3114  0.5082
#>   0.0949 -0.3464 -0.4862  0.0286  0.1515 -0.3816 -0.1246  0.1599  0.6792
#> 
#> Columns 10 to 18  0.2124 -0.1395  0.2149  0.4782  0.0698  0.4674  0.2134 -0.0375  0.0808
#>   0.2478  0.2917  0.0798  0.1362  0.4279  0.0980  0.2634 -0.5600 -0.1957
#>   0.6231  0.2835  0.0424  0.3099  0.4515  0.5818  0.0055 -0.1863 -0.2242
#> 
#> Columns 19 to 20  0.4261  0.0091
#>  -0.0710 -0.1945
#>   0.2094  0.1459
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>