RNN module

Applies a multi-layer Elman RNN with \(\tanh\) or \(\mbox{ReLU}\) non-linearity to an input sequence.

Usage

nn_rnn(
  input_size,
  hidden_size,
  num_layers = 1,
  nonlinearity = NULL,
  bias = TRUE,
  batch_first = FALSE,
  dropout = 0,
  bidirectional = FALSE,
  ...
)

Arguments

input_size

The number of expected features in the input x

hidden_size

The number of features in the hidden state h

num_layers

Number of recurrent layers. E.g., setting num_layers=2 would mean stacking two RNNs together to form a stacked RNN, with the second RNN taking in outputs of the first RNN and computing the final results. Default: 1

nonlinearity

The non-linearity to use. Can be either 'tanh' or 'relu'. Default: 'tanh'

bias

If FALSE, then the layer does not use bias weights b_ih and b_hh. Default: TRUE

batch_first

If TRUE, then the input and output tensors are provided as (batch, seq, feature). Default: FALSE

dropout

If non-zero, introduces a Dropout layer on the outputs of each RNN layer except the last layer, with dropout probability equal to dropout. Default: 0

bidirectional

If TRUE, becomes a bidirectional RNN. Default: FALSE

...

other arguments that can be passed to the super class.

Details

For each element in the input sequence, each layer computes the following function:

$$ h_t = \tanh(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh}) $$

where \(h_t\) is the hidden state at time t, \(x_t\) is the input at time t, and \(h_{(t-1)}\) is the hidden state of the previous layer at time t-1 or the initial hidden state at time 0. If nonlinearity is 'relu', then \(\mbox{ReLU}\) is used instead of \(\tanh\).

Inputs

input of shape (seq_len, batch, input_size): tensor containing the features of the input sequence. The input can also be a packed variable length sequence.

h_0 of shape (num_layers * num_directions, batch, hidden_size): tensor containing the initial hidden state for each element in the batch. Defaults to zero if not provided. If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Outputs

output of shape (seq_len, batch, num_directions * hidden_size): tensor containing the output features (h_t) from the last layer of the RNN, for each t. If a :class:nn_packed_sequence has been given as the input, the output will also be a packed sequence. For the unpacked case, the directions can be separated using output$view(seq_len, batch, num_directions, hidden_size), with forward and backward being direction 0 and 1 respectively. Similarly, the directions can be separated in the packed case.

h_n of shape (num_layers * num_directions, batch, hidden_size): tensor containing the hidden state for t = seq_len. Like output, the layers can be separated using h_n$view(num_layers, num_directions, batch, hidden_size).

Shape

Input1: \((L, N, H_{in})\) tensor containing input features where \(H_{in}=\mbox{input\_size}\) and L represents a sequence length.

Input2: \((S, N, H_{out})\) tensor containing the initial hidden state for each element in the batch. \(H_{out}=\mbox{hidden\_size}\) Defaults to zero if not provided. where \(S=\mbox{num\_layers} * \mbox{num\_directions}\) If the RNN is bidirectional, num_directions should be 2, else it should be 1.

Output1: \((L, N, H_{all})\) where \(H_{all}=\mbox{num\_directions} * \mbox{hidden\_size}\)

Output2: \((S, N, H_{out})\) tensor containing the next hidden state for each element in the batch

Attributes

weight_ih_l[k]: the learnable input-hidden weights of the k-th layer, of shape (hidden_size, input_size) for k = 0. Otherwise, the shape is (hidden_size, num_directions * hidden_size)

weight_hh_l[k]: the learnable hidden-hidden weights of the k-th layer, of shape (hidden_size, hidden_size)

bias_ih_l[k]: the learnable input-hidden bias of the k-th layer, of shape (hidden_size)

bias_hh_l[k]: the learnable hidden-hidden bias of the k-th layer, of shape (hidden_size)

Note

All the weights and biases are initialized from \(\mathcal{U}(-\sqrt{k}, \sqrt{k})\) where \(k = \frac{1}{\mbox{hidden\_size}}\)

Examples

if (torch_is_installed()) {
rnn <- nn_rnn(10, 20, 2)
input <- torch_randn(5, 3, 10)
h0 <- torch_randn(2, 3, 20)
rnn(input, h0)
}
#> [[1]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9 -0.8481  0.3023  0.1189  0.0194 -0.6420 -0.5907  0.3225  0.6136 -0.4085
#>  -0.6113 -0.8519 -0.2843 -0.9350  0.5404  0.6677 -0.2188  0.6350 -0.2148
#>   0.7629 -0.2264  0.7801 -0.8609 -0.7993 -0.7304 -0.1271 -0.0251 -0.2393
#> 
#> Columns 10 to 18 -0.3131  0.4389 -0.0124  0.2099 -0.4745  0.2528  0.3725  0.2471 -0.6662
#>  -0.0992 -0.6689  0.3791 -0.4809  0.9030 -0.3296  0.2911 -0.0896 -0.8896
#>  -0.5778 -0.3338 -0.2910 -0.5669 -0.0026  0.7032  0.8177  0.0398  0.1711
#> 
#> Columns 19 to 20 -0.3867  0.3960
#>   0.9243 -0.6892
#>   0.5982  0.4637
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4383  0.2305  0.1464 -0.0264 -0.4141  0.6190  0.2364  0.1007 -0.2465
#>  -0.6028 -0.4278 -0.1485  0.4118 -0.1175  0.0352 -0.0738  0.8844  0.4193
#>   0.1248  0.0004  0.2651  0.0465 -0.7187 -0.3614 -0.1437  0.4004 -0.0025
#> 
#> Columns 10 to 18 -0.5209  0.1247  0.7501  0.3769 -0.2055  0.3072  0.3332 -0.2942  0.3958
#>   0.2656  0.5575  0.0076 -0.4518 -0.3142 -0.2921 -0.3298  0.0471  0.3083
#>  -0.1088 -0.2272  0.0916  0.1070 -0.4684 -0.3001  0.1666 -0.5593 -0.5338
#> 
#> Columns 19 to 20  0.2203  0.1285
#>   0.1600  0.3761
#>  -0.1856 -0.0075
#> 
#> (3,.,.) = 
#>  Columns 1 to 9 -0.2171 -0.1144  0.0599 -0.2184 -0.4712  0.3689 -0.1281  0.3212 -0.0442
#>  -0.2432  0.0756  0.2432 -0.5378 -0.6957  0.3819  0.0042  0.5620  0.4163
#>   0.2066  0.1245  0.6560 -0.0560 -0.1388  0.2932 -0.1956  0.6479  0.3276
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{5,3,20} ][ grad_fn = <StackBackward0> ]
#> 
#> [[2]]
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 9  0.4737  0.6272 -0.3829  0.6375  0.4777 -0.6525  0.1707 -0.0293 -0.3988
#>   0.2632  0.5466 -0.3031  0.3242 -0.1856  0.0745  0.6754  0.0478 -0.4140
#>   0.7413 -0.1207 -0.4010  0.6811 -0.3735 -0.7478  0.1765 -0.6198  0.3394
#> 
#> Columns 10 to 18 -0.3309 -0.4449  0.2251 -0.0025  0.3601 -0.3336 -0.0107  0.3418 -0.0674
#>   0.6134  0.1191  0.0159 -0.1657  0.3992  0.5254  0.1860 -0.2419  0.5332
#>  -0.4287 -0.4018  0.3867  0.5516  0.4686 -0.1317 -0.0678  0.1858 -0.6784
#> 
#> Columns 19 to 20 -0.0049 -0.0072
#>   0.4130  0.2339
#>   0.0998  0.2222
#> 
#> (2,.,.) = 
#>  Columns 1 to 9 -0.4086 -0.3089  0.0567 -0.3018 -0.2731  0.5287 -0.1302  0.3550  0.2856
#>  -0.2195  0.0570 -0.2758 -0.2484 -0.1081  0.3818  0.4181  0.6483  0.4686
#>  -0.6665 -0.1345  0.3652  0.1872 -0.3369  0.3234  0.1456  0.5585  0.2982
#> 
#> Columns 10 to 18  0.4568 -0.0787  0.3831 -0.3764 -0.2411 -0.0885  0.1703 -0.3611 -0.1676
#>   0.1934 -0.1195  0.3529  0.1683 -0.6435  0.4519  0.1987 -0.5741 -0.2208
#>  -0.0138  0.2861  0.6502 -0.5594 -0.1924  0.2729  0.0880 -0.1832 -0.1388
#> 
#> Columns 19 to 20  0.1064  0.4190
#>   0.0092  0.3643
#>   0.6380  0.4917
#> [ CPUFloatType{2,3,20} ][ grad_fn = <StackBackward0> ]
#>