Gels

Computes the solution to the least squares and least norm problems for a full rank matrix A of size (m,n) and a matrix B of size (m,k).

tch_gels(x, A)

Arguments

x
tensor object

A
the m,b by nn matrix A

Examples

A <- tensor(matrix(runif(100), ncol = 10))
y <- tensor(matrix(runif(10), ncol = 1))
tch_gels(y, A)

#> [[1]]
#> tensor 
#>  1.0138
#>  0.4143
#>  1.6567
#> -0.4578
#> -0.1969
#> -0.1644
#>  0.1614
#> -0.0343
#> -0.6291
#> -0.8046
#> [ Variable[CPUDoubleType]{10,1} ]
#> 
#> [[2]]
#> tensor 
#> -1.2032 -1.6076 -1.6311 -1.3099 -1.3174 -1.3879 -0.9874 -1.1302 -1.6559 -1.0350
#>  0.3315 -1.1106 -0.5708 -0.2366 -0.4940 -0.0209 -0.5695 -0.5599 -0.8170 -0.5623
#>  0.0347  0.6010 -1.3241 -0.2729 -0.9134 -0.7683 -0.4830 -0.3341 -0.2006 -0.8785
#>  0.0017  0.0635  0.5244 -0.7463  0.2580 -0.0749 -0.0342 -0.0325  0.0466 -0.1460
#>  0.3366  0.0646 -0.0766  0.2492 -0.7881 -0.2053  0.2397 -0.6700 -0.4040  0.0730
#>  0.1655  0.1799  0.2899  0.4486 -0.6585  0.6353  0.2019  0.3352  0.2283  0.0848
#>  0.1736  0.2594 -0.0490 -0.0418 -0.1423  0.1200 -0.9275 -0.1349 -0.2672 -0.2452
#>  0.3750 -0.0147 -0.3837  0.2415  0.3968 -0.3187 -0.3924  0.2126 -0.3712  0.2498
#>  0.1708 -0.2066  0.2578 -0.2077  0.0588  0.4276  0.6337 -0.8683 -0.6660  0.7379
#>  0.4267 -0.2786 -0.2059 -0.1109  0.1293 -0.0127 -0.1242 -0.1323 -0.9826  0.1476
#> [ Variable[CPUDoubleType]{10,10} ]
#>