Baddbmm

Baddbmm

torch_baddbmm(self, batch1, batch2, beta = 1L, alpha = 1L)

Arguments

self
(Tensor) the tensor to be added

batch1
(Tensor) the first batch of matrices to be multiplied

batch2
(Tensor) the second batch of matrices to be multiplied

beta
(Number, optional) multiplier for input (\(\beta\))

alpha
(Number, optional) multiplier for \(\mbox{batch1} \mathbin{@} \mbox{batch2}\) (\(\alpha\))

baddbmm(input, batch1, batch2, *, beta=1, alpha=1, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices in batch1 and batch2. input is added to the final result.

batch1 and batch2 must be 3-D tensors each containing the same number of matrices.

If batch1 is a \((b \times n \times m)\) tensor, batch2 is a \((b \times m \times p)\) tensor, then input must be broadcastable with a \((b \times n \times p)\) tensor and out will be a \((b \times n \times p)\) tensor. Both alpha and beta mean the same as the scaling factors used in torch_addbmm.

$$ \mbox{out}_i = \beta\ \mbox{input}_i + \alpha\ (\mbox{batch1}_i \mathbin{@} \mbox{batch2}_i) $$ For inputs of type FloatTensor or DoubleTensor, arguments beta and alpha must be real numbers, otherwise they should be integers.

Examples

if (torch_is_installed()) {

M = torch_randn(c(10, 3, 5))
batch1 = torch_randn(c(10, 3, 4))
batch2 = torch_randn(c(10, 4, 5))
torch_baddbmm(M, batch1, batch2)
}

#> torch_tensor
#> (1,.,.) = 
#>   0.6709 -0.2889 -0.1143  1.2287  2.6792
#>   0.7328  6.8952 -2.1671  0.5258 -0.9268
#>   0.9975 -2.0659 -5.6112 -0.5098 -2.2135
#> 
#> (2,.,.) = 
#>   2.2435  0.1128 -1.2655 -2.8685  1.2882
#>   0.4596 -4.7321  1.0416 -1.8929  0.3076
#>  -0.8792 -0.6508  0.2442  1.4607 -1.7716
#> 
#> (3,.,.) = 
#>  -0.7823 -4.3495  5.7029 -3.8182  1.7295
#>   1.2738  1.2181 -4.3683 -0.2125  2.0777
#>   1.3731  1.9842 -2.2067  0.5831  0.0530
#> 
#> (4,.,.) = 
#>   2.8760 -1.9122  1.3080 -1.3768 -0.8532
#>   0.8094  1.7166 -0.9425 -1.1198  1.0302
#>  -5.0524 -0.5338  3.1819 -0.9377  2.1418
#> 
#> (5,.,.) = 
#>   2.1194 -1.3427 -1.3215 -2.4445 -4.0644
#>  -0.1634 -1.8138  4.1251  3.7300  1.0823
#>   1.0493 -5.8691 -2.6824 -1.8173 -5.8446
#> 
#> (6,.,.) = 
#>  -4.1280  5.6228  2.2389 -3.4616  2.0946
#>  -2.0582  1.2381  1.1488  0.9152 -0.0410
#>  -2.0075 -0.4832 -2.4516 -0.9541 -1.9924
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]