Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   2.9017  1.5114 -1.4097 -1.0548 -1.2422
#>   1.7102  1.8813  0.2205  2.0213  0.2034
#>  -2.0356 -2.6949 -0.0803 -2.7241  2.2157
#> 
#> (2,.,.) = 
#>  -0.8237  3.0374 -4.0096 -1.0299  3.5981
#>  -0.0752 -1.1889  0.6516 -0.2324 -3.9645
#>   0.2991 -0.9236  1.2980  0.3688 -1.2248
#> 
#> (3,.,.) = 
#>  -3.4203  0.5159 -0.1963  0.6452  0.2038
#>  -0.6345 -2.0553 -0.0044  6.3856  2.8342
#>   5.4447  1.8178  3.1374  5.7278  0.9118
#> 
#> (4,.,.) = 
#>   1.0893 -1.1285 -0.8921  0.8631  0.8043
#>   1.4391 -3.5273  0.7853 -0.5394  0.5270
#>   2.1121 -0.3716 -0.6091 -0.0946 -0.1780
#> 
#> (5,.,.) = 
#>  -1.5499  0.5419  0.3706  0.0724  0.4523
#>   1.8795  1.4978  1.4105  1.4917 -0.1104
#>  -0.8070 -1.5694 -2.0154 -1.4701 -0.4626
#> 
#> (6,.,.) = 
#>  -4.4089 -1.3425 -0.0593 -0.2796 -0.5227
#>   3.0329  3.6840 -4.5156 -1.2488 -1.0838
#>  -1.0311 -0.1022  1.4011 -0.1218 -0.0378
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]