Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>    1.5700  -2.0067  10.1988  -9.1316  -2.9252
#>    0.8770   4.3726  -3.6398   1.6664   1.1868
#>   -3.2809  -2.0775  -5.9366   2.5655  -2.2698
#> 
#> (2,.,.) = 
#>   1.9683 -0.7078 -1.4985 -0.8536  0.8929
#>   2.7925 -3.2069 -0.6893  0.5679  0.0008
#>   0.2257  4.4701  2.9387 -5.2397  3.0405
#> 
#> (3,.,.) = 
#>    2.8214  -0.0966  -1.9733   0.6823  -1.1191
#>   11.1402  -0.5802  -1.2605   1.8593  -2.6744
#>    1.1158  -1.1126  -3.4536   0.8224  -1.6829
#> 
#> (4,.,.) = 
#>  -1.7156 -0.1982 -0.3959 -2.0004 -0.8906
#>   1.6384 -1.3082 -1.9363  0.0556 -0.7476
#>   1.1065  0.7830  0.5882  5.9775  1.2405
#> 
#> (5,.,.) = 
#>   0.1510 -0.5524  1.1594 -0.1693  1.5997
#>  -2.9291 -0.2680 -0.2163  0.2881 -0.7121
#>  -3.3728  0.2017  0.7585  0.4115 -1.3098
#> 
#> (6,.,.) = 
#>   1.7582  1.6974 -0.5410  0.5853  1.8668
#>   4.1697 -0.4435  4.1606  0.7203 -0.5755
#>  -3.9136 -2.5512 -2.0472 -1.7891 -1.7694
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]