Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.4689  0.2981 -1.8392 -1.6875 -0.1244
#>   0.3360  0.2387  0.1897 -0.9903 -1.5107
#>  -0.7971  1.6493 -2.9451  0.5611  2.0878
#> 
#> (2,.,.) = 
#>  -0.3648 -0.4170  0.3581 -0.6307  0.3456
#>   1.9827  1.2922 -0.2769  1.3719  0.7723
#>   0.4310 -0.8174  0.4555 -2.0173  1.6790
#> 
#> (3,.,.) = 
#>   1.1588  6.0700 -5.8750  3.9089 -0.4183
#>   0.6908  2.7303 -1.2375  1.1816 -1.7898
#>   2.6958  3.5259 -1.6649  2.6406  2.5655
#> 
#> (4,.,.) = 
#>   0.0042  1.0651  0.6784 -1.7472 -6.5086
#>  -1.0202 -3.0131 -0.2537  0.7660  3.5814
#>  -4.4217 -1.5395  1.2061 -1.4706 -3.4410
#> 
#> (5,.,.) = 
#>   3.9392  2.1979  4.6072 -5.8609  1.3201
#>   3.2225  4.6444  1.2338 -2.5050  3.1130
#>   2.4534  0.1442  4.6749 -3.6966  1.0961
#> 
#> (6,.,.) = 
#>  -0.5271  0.1966  0.2516  1.1057  0.2772
#>   0.5379 -0.0317  0.2851  0.0491  0.0355
#>  -0.0301 -0.2912  0.4179  0.5350 -0.4048
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]