Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -7.1148  5.4270  1.9677  1.4881  3.0304
#>  -1.2014 -0.8402  0.8511 -0.0612  0.2259
#>   0.6716 -0.9408 -1.0888 -0.2046  0.3333
#> 
#> (2,.,.) = 
#>  -1.3325  1.1644  3.4708 -0.0854  0.5289
#>   0.1855 -0.8465 -2.5910 -1.5482 -0.4552
#>  -4.0936  0.4802  7.8465 -0.7024  1.0927
#> 
#> (3,.,.) = 
#>  -1.4532  6.0711 -0.8052 -0.8021 -5.3499
#>  -0.5898 -0.3623 -1.8846  1.4022 -0.4462
#>   0.6799 -6.4236  1.9620  2.4896  8.8940
#> 
#> (4,.,.) = 
#>  -1.3810  2.2829  0.2663  2.1256  3.6051
#>  -0.0947  0.2557  0.5288  1.3299  0.3349
#>  -0.0859  0.8530  0.7507  2.3757  1.1537
#> 
#> (5,.,.) = 
#>   3.3619 -1.2145  2.5437 -0.0836 -0.2037
#>   2.2026 -3.4698  0.8757  2.5008 -1.2921
#>  -2.9025 -1.7808 -1.3767 -2.3048  0.8849
#> 
#> (6,.,.) = 
#>   1.6882  1.2855 -1.1408  0.0628  0.2890
#>   1.6947  0.0918 -1.2762  0.7803 -0.0348
#>   3.2923  2.1300 -2.1232 -0.9158 -0.0241
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]