Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   2.5038  2.4046  0.3552  1.1990 -0.4939
#>  -1.3554  0.2856 -0.0867 -0.1279 -1.4805
#>   0.6516 -0.5331  0.2187  0.7422 -0.3755
#> 
#> (2,.,.) = 
#>  -2.6122 -0.5406 -0.5520  0.5588 -0.9676
#>   0.0973 -0.3184 -0.4978  1.0920 -0.8192
#>   1.0893 -0.5257  1.0724  1.0008  0.0035
#> 
#> (3,.,.) = 
#>  -3.4118  0.6171 -3.6242  3.0366  0.7318
#>  -3.9567 -0.1635 -2.3998  3.0755  1.1109
#>   1.1522  0.6764 -1.1339 -0.8247  0.3083
#> 
#> (4,.,.) = 
#>  -2.5650  0.7531 -0.8027  1.4477 -2.8821
#>  -2.7013 -1.2988  1.5392  2.8633  0.4156
#>  -0.8335 -2.0357 -0.2332 -0.8202 -0.1724
#> 
#> (5,.,.) = 
#>  -3.0474  1.7695  1.8519  0.7795 -1.8745
#>   0.2408 -1.5338 -1.5424 -1.4697 -0.1326
#>   0.1551  3.2238  4.0001  3.2001 -2.0167
#> 
#> (6,.,.) = 
#>   0.4490  1.0446 -1.8157 -1.6165 -1.0237
#>  -1.1870 -5.5903 -0.7162  0.6479 -5.9046
#>   1.4355 -0.7121 -0.6566 -0.6130 -2.5719
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]