Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -2.6492  0.7203  0.6470  1.0748  0.8508
#>   0.0058 -2.0582 -0.3531  0.1756  0.8454
#>  -1.3357 -2.4554 -1.7316 -1.8185  0.0013
#> 
#> (2,.,.) = 
#>  -2.7654 -1.7565  1.0467 -0.9021 -0.2101
#>   2.9694  2.6698 -1.5012 -0.1578 -2.3553
#>   0.0521 -0.5841  0.8336 -1.0225  0.9534
#> 
#> (3,.,.) = 
#>  -3.0198 -2.6526 -0.6884  2.7970  1.8561
#>   1.3988  1.6196  0.8264 -1.6880 -0.5836
#>  -1.8361 -1.6435 -2.5033  2.5535  1.9679
#> 
#> (4,.,.) = 
#>   0.1864 -0.7930 -0.9378 -0.4791 -0.9357
#>  -0.8391 -0.1271 -0.2845 -0.1921  0.2707
#>   1.0138  0.6315 -0.6457 -3.1668  0.2965
#> 
#> (5,.,.) = 
#>  -3.1936 -0.0065 -1.4337 -5.8173 -3.8842
#>   1.6228 -2.4077 -2.8693  2.7452  2.5736
#>   2.7376 -1.3444 -2.8333 -0.6168 -1.9218
#> 
#> (6,.,.) = 
#>   2.4491  1.3780  2.5484 -4.9873 -1.3377
#>   0.9654  0.1103  0.5957 -0.8606 -0.6628
#>   1.0585 -0.1270  0.3515 -0.7803 -1.9257
#> 
#> (7,.,.) = 
#>   1.2230 -0.9206 -0.4467 -4.8658  1.2931
#>   0.0427 -0.3020  0.3428 -2.8634  0.2969
#>  -1.4106  1.3028  0.5776  4.2019 -1.9976
#> 
#> (8,.,.) = 
#>  -3.3022  1.9207 -1.7976 -1.0289 -0.2126
#>  -0.1250 -2.7791  0.7749  0.8356  5.0383
#>   3.4486 -1.9389  5.0328 -0.6195  1.6185
#> 
#> (9,.,.) = 
#>   4.4574  2.1288 -3.2226  2.0862 -1.9960
#>   1.9326  2.3443 -0.1818 -0.1431  0.2726
#>   0.8138  1.2044  0.2947  0.2196  1.1001
#> 
#> (10,.,.) = 
#>  -0.3073 -2.6836  0.3500 -0.0734  1.5967
#>   0.9689  2.5241  1.7220 -0.3316  0.0733
#>   0.2914 -0.7167  6.0589  0.7247  3.6571
#> [ CPUFloatType{10,3,5} ]