Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   1.3449 -0.3518 -1.4911  1.1438 -0.5783
#>  -0.2759  1.4554 -0.0567  0.5219  1.3421
#>   1.8859 -0.7604  1.0628 -0.2322 -2.0097
#> 
#> (2,.,.) = 
#>   0.6587 -0.3145  0.6081  0.6649  0.0544
#>   3.3358 -0.0868  0.9935 -0.7265  3.8534
#>   1.4469  1.0394  0.5707  0.6834  1.8118
#> 
#> (3,.,.) = 
#>   0.5963 -0.2414  2.9776 -1.2678  1.7916
#>  -2.0017 -0.1008  0.4794  3.0001 -4.1093
#>   0.2491 -2.9235  2.7552  0.7396  0.8961
#> 
#> (4,.,.) = 
#>  -1.1614  0.5261  0.2331 -0.0265 -0.9409
#>   1.2746 -0.9461 -6.3384 -2.0119 -2.9281
#>  -1.5258  1.3625  3.6211 -1.0777 -0.2579
#> 
#> (5,.,.) = 
#>   0.3592 -0.4585 -1.9138 -1.6756 -0.6730
#>   2.0004 -1.5365  1.5337  2.4089  1.3687
#>   2.5728  3.8242  3.0039  0.6846  0.4405
#> 
#> (6,.,.) = 
#>   0.0061 -0.5308  0.1754 -1.2170 -0.5358
#>   1.3472  2.0564  1.1519  1.0249  1.3656
#>   0.6048  1.4114  0.2442  3.3664 -0.1782
#> 
#> (7,.,.) = 
#>   1.2009 -0.1827  2.3082  4.1859 -1.1510
#>   3.4778 -1.1772  0.3732 -1.0196 -0.5530
#>   3.8285 -0.4425  2.4566  2.9765 -0.4720
#> 
#> (8,.,.) = 
#>   0.7439 -0.0401 -1.3241  0.1724  0.1429
#>   1.0309 -0.3096 -0.5297  1.3026 -1.0188
#>  -0.3008 -0.5784 -1.2168  3.5255  0.0622
#> 
#> (9,.,.) = 
#>  -0.8183 -0.2000 -2.3357  0.3021  1.4646
#>  -2.3123  0.1103 -1.5779  3.6092  2.4286
#>  -2.6433  0.5089 -1.8097  4.7375  2.6299
#> 
#> (10,.,.) = 
#>  -4.4645  1.9183 -0.0929  2.1050 -0.1388
#>   6.7040 -2.8248  0.1570 -0.6528  3.1883
#>  -7.6036  2.0356 -1.4965 -4.0877 -1.6103
#> [ CPUFloatType{10,3,5} ]