Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   3.8416  1.0074 -2.4051 -3.6978  2.2922
#>  -1.1825 -5.0957  0.1359  5.9774  1.4903
#>   0.9599 -0.3698 -1.4809 -2.8463  1.5197
#> 
#> (2,.,.) = 
#>   1.2759 -0.1333 -1.8211  2.2640 -1.1603
#>  -3.8501 -0.3399 -0.0009  0.4305 -2.2840
#>  -0.0946  1.2576  1.4401  0.7336 -1.0687
#> 
#> (3,.,.) = 
#>  -0.0824 -3.5623 -2.7365 -4.5893  0.7048
#>  -0.3312 -0.0597  0.6652  1.2601 -1.1092
#>   0.2852 -0.8737 -0.8412 -2.9133  1.3581
#> 
#> (4,.,.) = 
#>   3.6612  1.7878  0.2143 -5.8075 -0.5973
#>   2.7842  2.2125 -0.3196 -4.1815  0.5701
#>   0.6252  0.7547 -0.0549 -2.1511 -0.1039
#> 
#> (5,.,.) = 
#>  -0.1796 -0.5235  3.3101 -4.7578  0.2906
#>  -0.4207  0.8034  0.8318 -2.4202 -0.5375
#>   1.1237  0.1732 -1.3715  2.5219  0.8718
#> 
#> (6,.,.) = 
#>   0.7963  0.9016 -1.2868 -0.4116 -0.4234
#>   2.7108  0.5328  0.0791 -0.0552 -0.5352
#>   2.8527  0.7764 -3.9340  2.2934  0.8463
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]