Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   1.0892  1.3748  0.8459 -2.0545 -1.4386
#>   0.3660 -0.1655  0.0205 -0.7982 -1.6220
#>   0.4613  0.7003  0.2010  1.5919  1.4269
#> 
#> (2,.,.) = 
#>  -2.2132 -1.0830  0.8680 -0.3013  0.5952
#>  -1.7194  1.7257 -1.8858  0.5295  0.0113
#>  -1.9067 -3.6883  3.4484 -0.6959  1.6596
#> 
#> (3,.,.) = 
#>   0.4246 -0.4291  0.5495  0.0139 -2.8828
#>   0.8869 -1.7168  0.9370  2.8693 -0.8925
#>  -0.1330 -0.6530  0.2961 -0.4661  0.2099
#> 
#> (4,.,.) = 
#>   3.2950  1.2318 -0.2358  3.7370  7.2076
#>   0.5021 -7.1274 -0.6974  2.2166 -1.0051
#>   0.5575 -0.9316 -0.4947  1.2719 -0.3476
#> 
#> (5,.,.) = 
#>  -0.7364 -1.5675  1.2250  0.6341  2.4889
#>  -3.3237  2.9684  3.2533  0.5785 -1.9284
#>  -3.7940  1.4235  3.7042 -0.9522 -1.3960
#> 
#> (6,.,.) = 
#>  -3.5668 -1.3443  5.1221 -1.1451  0.9565
#>  -1.3876  1.5025 -3.9358  1.3078  0.9739
#>  -1.9256 -2.0872  7.0016 -1.1434  0.6967
#> 
#> (7,.,.) = 
#>  -0.2430 -1.1612  0.2267 -0.4137  0.3275
#>  -1.0699  1.5163  1.1168  0.4205  1.8671
#>  -1.1855  5.5310  1.8558  4.2721  1.0853
#> 
#> (8,.,.) = 
#>  -1.0895 -2.0372 -2.1277  0.7614 -1.5783
#>  -0.6140  0.0763  1.5264 -0.7559 -0.3574
#>   0.2487 -1.0749  0.2471  1.2148  0.5383
#> 
#> (9,.,.) = 
#>  -0.7402  0.6009  1.3001  0.8246 -0.4576
#>  -0.1567 -0.6394  0.0932 -0.1400 -0.7153
#>  -0.5724 -0.2588 -1.4616 -1.0371 -0.5793
#> 
#> (10,.,.) = 
#>   7.8231  2.6005 -0.8253 -0.6163 -4.4321
#>   3.5758  0.7528  0.7691  3.3599 -1.3278
#>  -3.2284 -0.5208 -1.5541 -2.5006  2.0721
#> [ CPUFloatType{10,3,5} ]