Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.5967 -0.1546 -0.7700  1.0818 -0.1829
#>  -0.9639  0.1224 -1.5456  1.3503  1.2236
#>  -0.6344  0.2912  2.1134 -1.5930 -0.8024
#> 
#> (2,.,.) = 
#>  -0.0544  1.2392 -0.3264 -1.1663 -1.2571
#>  -0.5250  3.2617  0.2252 -3.6877 -1.3786
#>  -1.6709  1.3641  2.2251 -1.5906 -1.4930
#> 
#> (3,.,.) = 
#>  -2.6019  2.8372  0.9157 -2.6698 -0.8251
#>   0.2674 -2.0825  1.8705  0.7364 -3.6752
#>  -1.7407 -0.0823 -0.0812 -1.5332  1.5698
#> 
#> (4,.,.) = 
#>  -2.0226  2.6184  2.0043  3.7078  1.3961
#>   0.5983  0.3855  1.4825  0.8618  1.0958
#>   1.4246  0.5441 -0.0415  1.0105 -0.6502
#> 
#> (5,.,.) = 
#>   0.2020  3.3807  1.6334  1.2364 -3.1004
#>   2.4163 -0.9529 -3.6589 -0.3878  7.9713
#>   0.5421  1.0587 -1.8628 -0.3910  3.8588
#> 
#> (6,.,.) = 
#>  -0.8862  3.6379  2.1310  0.9314 -0.5842
#>  -2.1014  2.9959  1.9742  1.1552  5.9489
#>  -2.4885 -3.7498 -1.6245  0.4591  4.0684
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]