Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -4.8496 -3.4338  1.9085  6.6871 -3.7470
#>   2.1401  2.2811 -0.1298 -3.0487  1.9078
#>   5.5620 -0.4544  1.4762  0.3908  0.6415
#> 
#> (2,.,.) = 
#>   2.6299  0.2754 -2.5957  0.8264 -0.9926
#>  -0.4501  0.0284  0.4146 -0.0141  0.3347
#>   0.1395 -0.0361 -3.6140 -2.8806 -1.2973
#> 
#> (3,.,.) = 
#>   2.2255 -1.6704 -0.8549 -6.1328 -1.0277
#>  -1.8448  0.5652  0.4087  1.2867 -0.5934
#>   0.4394  0.2310 -0.2743 -0.5041 -1.7842
#> 
#> (4,.,.) = 
#>   0.2682 -2.2194 -0.2623  1.0628  0.3123
#>   0.5763  2.0047 -0.4135 -0.9159 -1.0987
#>  -1.8901  5.8435 -0.2116 -0.3200 -2.0450
#> 
#> (5,.,.) = 
#>   0.4616  1.2059 -0.0621 -0.9446  1.9300
#>  -2.0071 -7.4455  1.3265  2.5542 -4.3254
#>   0.2308 -0.8500 -0.3007 -0.0094  1.1263
#> 
#> (6,.,.) = 
#>   0.9592 -1.5106  0.8917  1.6345  1.6022
#>   1.1646  2.3546 -3.6881  0.0239 -1.5678
#>  -0.9073 -0.8794  0.4793 -0.4412 -1.9063
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]