Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   2.8837  1.7028  0.0612  0.0441 -0.8363
#>   0.1355 -0.0820 -0.3646  0.6502  0.2941
#>   0.4807  0.2395 -0.1419  1.1769  0.1385
#> 
#> (2,.,.) = 
#>  -0.2087  0.6893 -0.7844 -0.6650 -1.2998
#>  -3.2399  1.9847 -0.3897  0.3850 -0.0124
#>  -0.0163 -4.3319  1.2836  1.0473  2.4989
#> 
#> (3,.,.) = 
#>   2.4992  1.3188 -0.0828 -1.5224  1.6522
#>  -0.8417  0.5516  2.3586  1.7337  2.1697
#>  -1.4209  1.4740 -1.3640  0.1035  3.2141
#> 
#> (4,.,.) = 
#>   0.8895  2.3114  2.9946  0.2976 -1.9407
#>  -1.9199 -0.7722 -1.3826  0.4278 -0.5088
#>   1.8879 -0.4393 -1.4661 -1.1069  1.4349
#> 
#> (5,.,.) = 
#>   2.3660 -1.9215 -0.7735 -0.1746  0.6412
#>   0.9593  2.2068 -2.4433  0.3645 -0.0780
#>  -0.7772  0.6493  1.5485 -1.5039 -2.4614
#> 
#> (6,.,.) = 
#>  -0.9379  1.9031  0.4133  2.3471 -3.7086
#>  -0.9122 -1.4216  2.8674 -2.0632 -0.7597
#>   0.3498  0.7166  1.4316  3.7973 -6.2659
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]