Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.9037 -0.1973 -1.1341 -0.9858  0.3039
#>   0.2868  0.2786  0.3566 -0.7423 -2.7733
#>   5.4676 -3.4729  3.8975 -1.7444  0.4233
#> 
#> (2,.,.) = 
#>  -3.1970 -1.4906 -0.8008 -2.1444 -1.2875
#>  -0.9690 -1.4588 -0.1471 -1.2726 -0.6858
#>  -0.3189  0.5680 -0.3345 -0.9867 -0.8067
#> 
#> (3,.,.) = 
#>   0.3459  0.5981 -1.5873  0.8666 -0.4362
#>   0.2688 -1.3614  0.0694  1.3043  1.0296
#>   1.8249  1.2295 -0.2098 -0.7566  0.0848
#> 
#> (4,.,.) = 
#>  -0.6213  1.2877 -0.9306 -0.4368  2.4335
#>   3.6945  0.7509  2.2091  4.8111  8.8770
#>  -1.6086  1.6030 -0.6769 -3.3762 -0.2409
#> 
#> (5,.,.) = 
#>   2.7608 -0.6829  3.4384 -3.7808 -0.0867
#>  -0.0835 -0.5113 -0.2495  0.1662 -0.8862
#>   0.1986 -0.7221 -1.9765 -0.4381  1.4731
#> 
#> (6,.,.) = 
#>   1.3646  1.3529  1.7244 -0.7949 -2.1482
#>   0.6172  0.6641 -0.1706 -1.2928  2.1358
#>  -0.5524  0.0898 -1.1729  2.1522 -1.7189
#> 
#> (7,.,.) = 
#>   1.3161 -1.3085 -0.4967  2.2483  0.0628
#>  -2.1286  2.2143  1.0357 -2.5808  0.1258
#>  -0.2157  1.8414  0.0169 -2.9079 -0.6500
#> 
#> (8,.,.) = 
#>  -1.4650  1.0743 -4.3104 -2.9395  1.6595
#>   2.2130 -0.8981 -1.0598 -1.1889 -6.1445
#>  -1.5164  0.6151 -3.6034 -1.3640  0.3356
#> 
#> (9,.,.) = 
#>   0.9710 -1.1168  0.7513 -1.7730  1.5852
#>   0.1585 -2.1137 -1.4338  0.9732 -0.3186
#>   0.9655 -2.5352 -5.2346  3.6612 -2.9682
#> 
#> (10,.,.) = 
#>   2.1552  6.8907  4.4197 -0.9053 -1.2060
#>  -0.7011 -3.1555 -4.6818 -0.2381 -1.1970
#>  -1.1343 -4.0314 -3.4194  0.7506 -1.1217
#> [ CPUFloatType{10,3,5} ]