Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -3.7946  1.4372  0.2643  0.5153 -1.0549
#>  -5.6417 -0.3039 -0.6462 -3.5776 -0.7586
#>   0.1720  1.1994 -3.0888  0.7061 -3.1860
#> 
#> (2,.,.) = 
#>   0.4602  2.4965 -1.2522 -2.7431 -2.5588
#>  -0.7037 -0.9544  1.0073 -0.9169  0.3563
#>   2.2417 -1.1718  0.3158  1.0289  0.1509
#> 
#> (3,.,.) = 
#>   0.6736  3.5190 -2.7650 -0.6781 -3.2257
#>  -0.7716 -1.3326  2.9409 -1.1093  2.2059
#>  -2.1707  1.3736  0.8085  0.6613 -1.5013
#> 
#> (4,.,.) = 
#>   2.0445 -1.3074  1.3973  3.0332 -2.7969
#>  -3.4920 -0.5446  0.0419 -0.9028  3.1314
#>  -3.2155  1.2464 -0.3691 -0.4430 -0.3392
#> 
#> (5,.,.) = 
#>   0.3839 -0.3477 -1.6705 -1.2454 -1.0226
#>  -0.3133 -1.4464  1.5966  0.0336  1.3573
#>  -0.6605 -1.6487  1.8598  0.8163  1.8861
#> 
#> (6,.,.) = 
#>  -6.2769 -0.5307  2.7242  1.4786 -3.0735
#>   3.8730 -0.9136 -1.6038 -0.4646  1.6587
#>   1.3614  0.0401 -0.3969 -0.1581  0.4810
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]