Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -1.0971  1.1239 -0.3342  0.4245 -0.5502
#>  -0.7503  0.5804 -0.1745  1.5026  0.3864
#>   0.6486 -1.3513  0.4366 -2.3338  0.5793
#> 
#> (2,.,.) = 
#>  -1.8414 -3.6640 -1.6519 -0.3658  2.9565
#>  -0.2676  1.6023  2.0980 -0.1450 -2.7073
#>  -1.8810 -0.4315  3.3451 -0.4842 -3.3178
#> 
#> (3,.,.) = 
#>  -0.4020  1.1710  2.3879 -1.0983 -1.8879
#>   1.5179  1.7694  1.7397 -1.1173  0.1906
#>   0.8913  2.4348 -0.4518 -6.5519 -1.1885
#> 
#> (4,.,.) = 
#>  -0.6256  1.4457  0.3205 -1.2908 -1.3411
#>  -0.5525  2.1117  0.7645 -0.7843 -0.7864
#>  -0.1200  0.0571  1.0487 -0.8781 -0.2346
#> 
#> (5,.,.) = 
#>   0.2550  1.1473 -0.1517  1.2602  2.3772
#>  -0.1856 -0.8325  0.1376 -2.0206 -3.0048
#>  -0.7421 -1.6780  1.0094 -1.5136 -2.4469
#> 
#> (6,.,.) = 
#>  -1.9460  1.0920  0.5151 -1.1407 -1.7146
#>  -0.2544 -0.1463  1.8073  0.0063  0.4097
#>  -1.3321  0.4989 -1.5495 -1.0574 -1.9061
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]