Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   1.2853  0.1773  1.6439 -0.7276 -0.6614
#>   0.9144 -1.6999  0.8352 -0.3726  0.4595
#>   0.7720  2.4428  3.2431 -2.5719  2.1885
#> 
#> (2,.,.) = 
#>   1.8374  3.2930 -0.5793  2.5701  0.2614
#>   1.4772  3.8947 -2.1816  1.6790  1.0726
#>   0.2510  1.4837 -1.6514 -0.5868  2.4410
#> 
#> (3,.,.) = 
#>  -3.5216  0.7271 -1.8321 -0.9815  2.5039
#>  -2.6512  2.3344 -0.4942 -2.3436  0.3108
#>   0.2220  0.4075  1.4403  0.2646 -0.9450
#> 
#> (4,.,.) = 
#>  -1.5522 -2.8274  1.3097  0.9660 -1.3499
#>  -1.5898  0.8632  0.7365 -2.5904 -1.1270
#>   3.1609  2.3690  2.0197  0.4693  0.1715
#> 
#> (5,.,.) = 
#>   0.4839 -2.2074 -1.3084 -1.6152 -0.5670
#>   2.1218  0.0352 -0.3592 -0.7460  0.8173
#>   0.5632  1.8245  1.4707 -0.3534  0.6185
#> 
#> (6,.,.) = 
#>   2.2036  1.2816 -0.6867 -4.8853 -0.5735
#>   0.2739 -0.2358  0.3178 -0.3767  1.2391
#>  -1.2995  0.0271  0.3211  2.2204 -0.1865
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]