Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.2643 -0.7212 -0.7675 -3.0916 -3.3646
#>  -0.0958  1.4462 -0.4187  1.1626  0.6700
#>  -0.8537  1.6333 -0.7339  2.4873  1.2475
#> 
#> (2,.,.) = 
#>  -2.5551 -0.5604 -1.0623  0.4415 -1.7874
#>  -2.6988 -0.0618 -0.8888  0.3276 -2.2117
#>   2.4900  1.2068  0.3057  2.6764 -2.1324
#> 
#> (3,.,.) = 
#>  -2.3188  0.3057  0.7345  3.4135 -0.3126
#>  -2.2443  1.4742 -0.3925  0.4214 -0.6046
#>   4.4081 -1.8309  0.1354 -0.9060 -0.8332
#> 
#> (4,.,.) = 
#>   1.7682 -0.7963 -2.7066  0.0583  1.2486
#>   1.0862  1.3184 -2.2325  0.0554  1.9354
#>  -0.3897 -1.8641 -2.3145  1.0174 -0.9640
#> 
#> (5,.,.) = 
#>   0.1647  0.4206 -0.1066  0.5020 -1.1059
#>   0.8882  0.8006  3.0366 -2.3724 -1.0676
#>  -1.0691 -1.0848 -1.6741  0.9894  1.0812
#> 
#> (6,.,.) = 
#>  -0.7830 -0.2738  0.9522  1.5143 -2.2802
#>  -0.4785 -1.2616  1.7579 -0.4661 -0.7053
#>  -0.1390  1.6425 -0.8014 -0.8028  0.7127
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]