Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -2.7875 -0.8042 -0.2377  2.7254  1.2324
#>  -1.3179 -1.5640 -0.4280  0.9017  2.7954
#>  -5.8519 -4.0669  0.0944  1.7432  8.7395
#> 
#> (2,.,.) = 
#>  -0.1239 -0.2303  1.6848  1.2726 -0.6816
#>   2.0503 -0.7523 -1.0496 -0.8244 -2.9871
#>  -1.6782 -1.4602  1.7135  0.1317  1.7523
#> 
#> (3,.,.) = 
#>   2.8947 -0.0164 -1.4802  1.1507  2.6518
#>   3.2898  1.1687  2.2244 -4.2655  3.9258
#>  -1.3899  0.8473  0.2474 -0.8773 -1.3137
#> 
#> (4,.,.) = 
#>  -0.8229  2.1365  0.6418  0.1636  1.2144
#>   2.0595 -0.6448 -0.7663  0.4084 -0.5527
#>   1.2704  2.9820 -1.7953 -1.2325  2.3860
#> 
#> (5,.,.) = 
#>  -1.8917 -1.1720 -2.3909 -1.0130 -3.1236
#>   0.0313  0.4113  0.2364  0.2581  0.7808
#>  -2.1091  0.7946 -1.8163  0.5738  0.9403
#> 
#> (6,.,.) = 
#>   0.0955 -0.2490 -2.1501  0.5041  1.7181
#>  -0.5084  3.8230  2.5337 -1.0699 -2.5486
#>  -0.1270  0.9051  0.0557 -0.0678 -0.1643
#> 
#> (7,.,.) = 
#>  -0.2927 -1.2679 -1.7212  2.6149  2.0771
#>   0.1421  1.4437  0.2179  3.9827  1.4667
#>  -0.1900  1.8258 -1.1202  0.6383 -0.9558
#> 
#> (8,.,.) = 
#>   0.6504  2.6118 -0.7319  0.2118  0.3674
#>   1.7915  2.9159  2.3021  3.2854 -2.8105
#>  -0.4084 -0.1065 -0.4272 -0.3426  0.3406
#> 
#> (9,.,.) = 
#>  -3.6094  2.5831 -3.1118 -0.9397 -1.4573
#>  -0.8359 -0.6007  0.6510  0.1193 -0.5118
#>  -0.7282  4.9849 -2.5798 -0.0552  3.9324
#> 
#> (10,.,.) = 
#>  -3.0691  5.1150  0.2371  2.4900 -2.6497
#>  -0.3790  2.2784 -2.7748 -3.0566 -1.6830
#>   2.5030 -4.3232  2.0600 -7.1850  2.3846
#> [ CPUFloatType{10,3,5} ]