Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.4455  0.5555 -0.1111  1.1703  0.5968
#>  -1.3731  0.9324  0.8594 -0.2582  2.2304
#>  -2.0484  2.3085 -1.1238 -0.0309  2.3635
#> 
#> (2,.,.) = 
#>  -0.5828  0.3749  0.7041  0.8747 -1.4563
#>   2.2398 -0.3050 -1.1984 -1.5242  3.8797
#>   5.8015  4.9510  4.1373  4.9269  1.4457
#> 
#> (3,.,.) = 
#>   0.8479  1.9389 -1.5220 -0.6965  1.5332
#>  -1.8539  0.2496  3.1334 -0.3984 -2.7092
#>  -2.8404 -1.4817 -0.2872 -2.6050  0.2247
#> 
#> (4,.,.) = 
#>   1.0975  1.2481 -1.0708  1.4758 -4.0178
#>  -0.0131 -2.3031  2.4727 -1.4560  1.1454
#>   0.2475 -1.4515  1.2536  1.1968 -0.9441
#> 
#> (5,.,.) = 
#>  -1.9292 -1.7373 -1.7411  1.2996 -6.3393
#>   2.5562  1.4616 -2.7792 -0.2685  1.2343
#>   0.5136  0.5959 -0.5252  0.3265 -0.5400
#> 
#> (6,.,.) = 
#>  -3.6554 -1.8041 -1.8663  2.8929  0.4173
#>   4.9698  1.4147  1.7442 -1.6061 -0.7482
#>   4.5106 -0.4561  0.6907  0.0486  1.4390
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]