Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   3.6611 -0.2426  0.8118  1.2438  2.1243
#>   0.6982  0.2105  0.6069 -0.7903 -0.6249
#>  -1.8525  0.2347 -0.2775  0.5018 -0.5187
#> 
#> (2,.,.) = 
#>  -1.1804 -5.0388  3.0964 -3.2161 -0.3063
#>   0.8087  0.8620 -2.4895  0.3953  0.1608
#>   1.4745  7.0362 -4.9551  2.8938  0.7438
#> 
#> (3,.,.) = 
#>   3.3308 -0.3547 -3.6500  2.3419 -0.6068
#>   7.4268  3.8023 -5.7414 -1.2369  0.4317
#>   0.3971 -0.6506 -0.3440 -0.3780  0.1325
#> 
#> (4,.,.) = 
#>  -0.8944 -2.8084 -0.1039 -0.6215 -0.1913
#>   0.7383  2.3866  2.4941 -0.7621 -2.9666
#>   1.6001  1.2770  0.1315 -0.0744 -0.7450
#> 
#> (5,.,.) = 
#>   2.2796  0.3255 -1.2226  0.7476  0.7834
#>   2.5994 -0.0796 -0.4931  1.3597  1.0759
#>  -0.9914  0.6675 -1.0259 -1.5234 -1.5779
#> 
#> (6,.,.) = 
#>   1.7483 -3.4226 -0.5540 -2.7475 -0.9348
#>  -0.1199  0.1175  1.2213  0.4493  1.0212
#>   1.4374  0.9787 -0.3007 -0.1427 -0.4918
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]