Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   2.1170  4.5270 -2.5701 -3.0572 -0.0087
#>  -0.3037 -1.2161 -1.2941 -1.1758  1.4827
#>   0.3534 -0.7214 -1.6887 -0.9797 -1.0249
#> 
#> (2,.,.) = 
#>   3.2956  4.5675  2.4555  1.2892  0.4661
#>  -0.7744  2.0691  1.6433 -0.9843  0.6042
#>   0.3444 -0.3169  0.2144  0.8315  0.6608
#> 
#> (3,.,.) = 
#>  -0.1997  0.8271  1.5514 -1.8728 -1.2764
#>   0.9386  3.0285 -0.1798  0.6027  0.4799
#>  -0.0291  0.7472 -1.0596  1.6153  0.7376
#> 
#> (4,.,.) = 
#>   0.4997  0.6820  0.0355  5.9989 -1.7865
#>  -1.0693 -1.4031  0.4766 -1.6215 -0.4046
#>   0.0793 -0.2244 -0.4865  1.9247 -2.1474
#> 
#> (5,.,.) = 
#>  -1.1281  0.6506 -3.1300  0.0877 -4.3085
#>  -0.6204  2.3490  6.2897  3.7338 -1.9572
#>   1.9833 -0.5227  8.0359  1.5937  2.0068
#> 
#> (6,.,.) = 
#>  -1.3653 -0.4918 -0.8464  0.3102 -1.3635
#>   1.0482 -0.0157  0.9653  2.7149  0.7985
#>   3.3059 -1.5284 -1.0656  0.2441 -1.0702
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]