Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.0428  2.9590 -1.7621  4.1941  4.4636
#>  -0.2932  1.1815 -0.9328  3.1043  2.3960
#>   1.0792  5.1097 -0.4560  1.2106 -0.4357
#> 
#> (2,.,.) = 
#>   1.9752 -0.1501 -0.3603 -0.0553  1.3744
#>   9.4771 -1.7073  4.3581 -0.2343 -1.3731
#>   0.6523 -0.0423  1.9464  0.3782 -0.3387
#> 
#> (3,.,.) = 
#>  -3.7480  2.2488  0.6737  0.0903 -2.5352
#>   1.7964 -2.7159  1.0151  0.6771 -1.6867
#>  -0.6470 -4.5101 -3.3614  0.9423 -1.7960
#> 
#> (4,.,.) = 
#>  -0.9385 -1.6936 -1.5189  3.3401 -1.7534
#>   1.2859 -1.6153 -3.8985  2.3107  1.0283
#>   0.9967  2.3699  0.8622 -1.7100  0.4445
#> 
#> (5,.,.) = 
#>  -0.1529 -0.5811 -0.2947 -0.4892  0.3960
#>  -0.7288 -5.2458  0.6194 -1.8301 -1.0493
#>   2.8957  0.9941 -3.4074 -0.8907  3.5075
#> 
#> (6,.,.) = 
#>   2.1297 -0.0910  1.9182 -0.5781 -2.1325
#>  -0.1019  1.4791 -2.3876  2.3543 -0.0488
#>  -0.1481 -1.1050 -1.6057 -0.3355  0.7515
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]