Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.2702 -1.4767  0.0047  1.4934  6.1982
#>   0.7180 -0.6899 -1.4310 -2.1534  0.0918
#>  -0.0356  3.2637  0.4643  1.5754  0.9000
#> 
#> (2,.,.) = 
#>   2.3470 -1.2556  0.3049  0.4641  0.6411
#>  -0.6555  1.0109 -0.3639  0.4662 -0.4285
#>   1.2452 -2.4032  0.1077 -1.1824  1.5148
#> 
#> (3,.,.) = 
#>  -1.1546  0.0950 -0.9633  0.5038 -0.9580
#>   3.9184 -2.3883 -3.2398  1.2722 -0.6353
#>   4.0142 -3.1632 -5.1138  0.9851  3.3117
#> 
#> (4,.,.) = 
#>  -2.4378  0.0713 -1.8055  0.6524  2.6658
#>  -0.5128  0.5415  1.1531 -2.7571 -2.7312
#>   2.3084 -0.9179 -0.7729 -1.2158 -0.2264
#> 
#> (5,.,.) = 
#>   0.5065  0.4423 -0.5080  0.9485  0.4276
#>  -0.4041 -0.3857  0.1177  0.0059 -0.0992
#>  -1.9611 -1.7425 -0.7590 -1.4488 -0.9653
#> 
#> (6,.,.) = 
#>  -1.6783 -2.5345 -3.7017 -4.0728  3.4991
#>   2.7059  2.0041  0.5661  3.1804  0.3722
#>  -1.9054  1.3023  4.9830 -1.5925 -0.1358
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]