Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.2247 -0.0408 -1.4708 -1.1981 -0.0863
#>  -1.3675  1.8516  0.5723 -1.8247 -1.7090
#>   0.8606 -0.2096 -0.2662 -1.7323  1.3148
#> 
#> (2,.,.) = 
#>   2.8217  0.1332  0.7727 -2.4131 -1.7174
#>  -6.3553  1.1083 -2.0074  1.6274  1.4953
#>  -1.8495  0.4749 -1.9065  0.5896 -0.8311
#> 
#> (3,.,.) = 
#>   3.3452  2.3748  0.5628  0.7036 -1.2459
#>   4.9033  2.8805  1.3605  0.3105 -2.9620
#>  -0.4337 -0.6321 -1.9610  0.2172  1.6325
#> 
#> (4,.,.) = 
#>  -0.8786  3.7101  1.6191  2.6022  3.4582
#>  -1.7835 -2.6029  7.7082 -2.1915 -4.6582
#>  -0.6293  0.3740 -3.6953  2.5622 -0.1144
#> 
#> (5,.,.) = 
#>  -5.3322 -1.7985 -0.7471  2.6454 -8.1035
#>  -0.3361  5.1678 -2.2245  4.0117 -1.9583
#>  -1.2032 -2.5194  1.2736 -0.1824 -1.8202
#> 
#> (6,.,.) = 
#>   0.6008 -3.2342 -0.0909  3.7705 -3.0094
#>   6.3176  4.4342  0.0735  1.2826 -0.5184
#>  -0.0204  2.8375  0.3073  0.9319  3.5348
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]