Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -1.6364 -0.8998  1.3097  2.5117  1.4074
#>  -0.1405 -0.0527 -1.0348 -0.8161  0.0891
#>  -0.5005  0.3738 -0.1189 -0.0823 -0.0794
#> 
#> (2,.,.) = 
#>  -0.4301  1.3900 -1.2051  0.5317  0.1110
#>  -1.6617  3.1434  1.7331 -0.2839  0.0818
#>  -0.9321  0.5416  0.3631 -0.2644 -0.7366
#> 
#> (3,.,.) = 
#>   1.3550 -1.1881 -4.3215 -2.0730 -0.1623
#>   0.2037 -0.4314 -0.4154 -0.4595  0.4852
#>  -0.3392  1.7774  3.6945  1.8783 -1.7801
#> 
#> (4,.,.) = 
#>   1.4355 -2.0163 -0.6767 -1.7228  1.4351
#>  -0.3392  1.5964 -0.9144  2.2464 -0.3602
#>   0.3437 -0.3769 -0.9789  1.4363 -1.8452
#> 
#> (5,.,.) = 
#>  -1.6381  1.6501 -1.2845 -0.8984 -0.5960
#>  -4.4053  4.1042 -1.4040 -4.9479  2.1211
#>  -0.4259  1.5205 -0.8797 -2.3037  0.6727
#> 
#> (6,.,.) = 
#>   0.2352  1.0883 -4.8700 -2.8125 -0.9836
#>  -1.6893  1.0306  0.7122 -0.5176  4.0851
#>  -1.1093  0.2666 -0.5027  0.2346  1.0949
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]