Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   1.0712  0.0486 -0.8446  0.9280 -0.9457
#>  -0.2525  1.3951  1.4122  0.6270 -0.7162
#>   1.6160  1.3178  0.9749 -2.3007  0.7046
#> 
#> (2,.,.) = 
#>   1.1110 -0.8926 -2.9785  2.3635  1.0275
#>   2.5889 -2.5455 -3.0315 -1.7638  1.9454
#>  -0.4920  0.4497  1.0265  1.7576  0.4888
#> 
#> (3,.,.) = 
#>  -1.3035 -0.6683 -0.6296 -0.1009  1.3188
#>  -0.2973 -2.3671 -0.0381  1.5464  0.7169
#>   0.9024  1.0602  0.3417  0.9734 -2.3594
#> 
#> (4,.,.) = 
#>   1.4470 -2.0335  0.0473 -2.2083 -4.3517
#>   0.3270 -0.1371 -1.7386 -0.7966 -0.0125
#>  -1.9626  1.3059 -3.4652 -0.9281 -0.3006
#> 
#> (5,.,.) = 
#>  -2.0953  2.6550 -1.0922  2.9563  0.2281
#>  -1.8391  1.1696 -1.1929  2.5825 -1.6397
#>   0.8258  2.1040 -1.7477  2.0769 -3.6952
#> 
#> (6,.,.) = 
#>   0.5758 -0.2771  5.2629 -4.5981 -1.0522
#>   0.4636 -0.6288 -0.1985  0.4217 -0.2346
#>   0.3499  0.3132 -1.0200  1.1688 -0.0524
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]