Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.3981  2.6049 -1.9652 -0.6496 -0.1257
#>  -0.0906 -1.8827  1.2410  0.9295  1.2941
#>  -0.0063  1.7605 -1.3300 -0.1329 -0.0136
#> 
#> (2,.,.) = 
#>   1.6336 -0.0863 -0.4173  0.6592  0.7676
#>   1.3649  1.5938 -0.2032 -4.1876 -1.8749
#>  -2.4909 -0.5911  0.1283 -3.9802 -0.3718
#> 
#> (3,.,.) = 
#>  -1.1502 -0.2050  1.2060 -2.2189  0.4321
#>  -3.8548 -0.2550  1.7597 -4.0009 -0.4822
#>   0.2109  2.7316  1.6480 -0.9617  1.5122
#> 
#> (4,.,.) = 
#>  -2.5098 -3.0846 -1.3896  1.4023  1.9131
#>  -2.4987 -2.0154 -1.5408  1.6152 -1.4945
#>  -1.6276 -0.6634 -1.8158  0.3781 -3.0603
#> 
#> (5,.,.) = 
#>   1.4066  0.1387  1.3478 -0.5555  1.3344
#>   2.9757  0.4744 -0.9452  0.5288 -1.9651
#>  -2.3792 -2.3759 -1.6100  2.3583 -0.2453
#> 
#> (6,.,.) = 
#>  -0.9753  2.6891 -1.0007 -2.8593  0.2835
#>  -3.4614 -0.1262  4.8742 -2.0294  0.0966
#>  -2.1551 -1.5665  4.3391  1.1281 -0.3545
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]