Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   1.2403 -0.1667  0.3666 -1.1507 -1.7299
#>  -0.2486 -0.7204 -0.4022  1.0860  0.0432
#>  -1.2908 -1.1179 -0.5473 -0.8211  2.5233
#> 
#> (2,.,.) = 
#>   0.5995 -0.7814  0.8130 -1.2709 -0.9134
#>  -0.7321  3.0291 -0.9954  1.1880  0.9003
#>  -1.4085  2.3214 -0.8535  1.4181  1.3118
#> 
#> (3,.,.) = 
#>  -1.6213  0.9274 -1.0721 -0.8786 -1.5182
#>  -0.8082  1.0572 -0.3072 -1.1384 -1.2536
#>   0.8572 -1.2918 -1.9431 -0.2820  0.4731
#> 
#> (4,.,.) = 
#>  -1.8322  0.1749  1.6403  2.8281  0.7929
#>   0.8667 -0.2166 -0.5428 -2.5746 -1.1575
#>   0.1874 -0.2158  0.3718  0.5170  0.1441
#> 
#> (5,.,.) = 
#>  -2.9918 -0.5094  0.9437  1.4280  3.4864
#>   2.9672 -0.7405  0.8885 -2.5544  0.2560
#>  -2.3135  0.9009 -3.9216  3.0711 -4.8937
#> 
#> (6,.,.) = 
#>  -0.1293 -0.4441  1.5963  0.9382 -4.4954
#>  -0.2850  1.8300  1.2245  1.0370 -2.4681
#>  -0.2307  2.7886  0.4221  0.6871  0.1367
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]