Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -2.0292  2.4826 -0.5474  0.6800  3.8577
#>   1.6635 -2.4374  1.0984  0.8035 -2.2743
#>  -0.7444  0.5562 -1.5810 -2.2048 -1.9256
#> 
#> (2,.,.) = 
#>   0.7864  1.8093 -0.0513  0.4432 -0.9157
#>  -0.6899 -1.1173  0.8012 -0.8515 -1.6592
#>   1.4628 -2.5764  2.3928  1.3369 -2.4060
#> 
#> (3,.,.) = 
#>   0.7788 -0.7743 -2.3777 -1.2461  0.0508
#>   0.3223  0.3742  2.4616  1.6609 -1.5246
#>  -0.7235  0.2993  2.0546  1.3558 -0.4408
#> 
#> (4,.,.) = 
#>  -2.2619 -2.3186  1.4302  0.4554  0.9529
#>   2.4746  4.0577  2.3066 -1.9608 -1.1304
#>  -0.9671  0.0455 -2.1332 -0.1224 -1.3329
#> 
#> (5,.,.) = 
#>  -0.9344 -2.1838  1.1080  1.5353  1.5097
#>  -2.2984 -3.2465  3.7871 -0.4786  0.0627
#>   0.9561  0.2035 -1.8129 -0.5077  0.6427
#> 
#> (6,.,.) = 
#>  -0.3233 -1.8181 -1.4841 -0.3113  1.7060
#>   0.7236 -0.3367  1.0951  0.8434 -1.7218
#>   1.3859 -0.9207 -0.9951  2.1247  0.4670
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]