Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.8425 -2.5332 -0.6754  0.9235  2.5688
#>  -0.2028 -2.8650  0.5843 -0.7103  1.6599
#>   1.4212 -2.4718  0.5928 -2.2154  3.1403
#> 
#> (2,.,.) = 
#>   7.0343 -2.0735  2.6108  5.1387  0.2675
#>   2.0126 -1.2328  0.5124 -0.0096 -2.5254
#>   3.4660  0.9859  2.4151  2.2084  0.1381
#> 
#> (3,.,.) = 
#>  -2.5702 -0.2326  0.7213  0.3191  2.3673
#>   1.1618 -0.2811  3.2600 -1.1805  0.9491
#>   3.5157  1.8912 -0.8301 -1.0196 -1.5494
#> 
#> (4,.,.) = 
#>   5.5452 -2.0610  0.5680  1.0294  1.0216
#>  -0.4404 -1.6060 -0.2815 -2.0042 -0.8539
#>  -2.2531  1.0674 -0.8160 -0.4840 -1.1171
#> 
#> (5,.,.) = 
#>  -1.1847  0.5301  1.5975 -1.1720  1.1681
#>  -0.8556  0.9508  0.5413 -1.7311  2.1473
#>   0.0018  0.6368  0.3484 -1.3654  0.1691
#> 
#> (6,.,.) = 
#>  -1.1556 -0.0154  1.0153  0.7737  0.1079
#>  -0.5306  0.4057 -0.0594  0.3116 -0.2600
#>   0.9429 -1.8072  0.2695 -0.7233 -0.5186
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]