Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.9057  0.0279  3.3535  1.1560  0.0070
#>   0.8092  1.5399 -0.6966 -3.5432 -0.4556
#>  -0.0870  1.0077  3.3688  0.1288  0.4654
#> 
#> (2,.,.) = 
#>  -2.3104  2.4775  0.4179  3.8787 -4.6113
#>   1.9398  1.2500  1.0341  1.6959  0.4538
#>   2.2621 -1.9204  1.1594  0.6962  1.4346
#> 
#> (3,.,.) = 
#>   0.2005 -3.1079 -0.5426 -3.1659 -1.3608
#>   0.1030 -0.9237 -2.6836  0.8760  1.0845
#>   0.3936 -3.3972 -4.2817  2.0883  2.1790
#> 
#> (4,.,.) = 
#>  -0.5846 -5.6843 -1.1138 -1.0093  3.7408
#>  -2.1245 -1.0460  0.6667  1.0093  2.6375
#>   0.2618  0.6783 -0.4628 -0.0592 -3.4235
#> 
#> (5,.,.) = 
#>   3.0883 -2.6216 -2.9948  0.7832  0.1870
#>   0.6945  3.4564 -1.3700  1.1381 -0.6264
#>   0.3207 -3.5854 -0.2068 -0.3943 -0.4608
#> 
#> (6,.,.) = 
#>  -1.5879 -0.2075  0.2025  6.2099  0.2199
#>  -0.6649 -2.9911  1.2124  5.3872  1.2823
#>   1.0333 -1.6585 -0.4312 -0.3422  0.0761
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]