Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -3.6054 -4.5753 -0.2474 -3.7817 -4.6050
#>  -1.1990  0.4792  0.1921  0.9700  0.9828
#>   2.3853  0.9318  1.3854  4.7278  6.7760
#> 
#> (2,.,.) = 
#>   1.1926  0.9949  1.9992  0.1926  1.8772
#>  -2.2442 -0.7078 -0.9200  1.2093  0.5770
#>  -1.6907 -1.9514 -1.1894  0.0670 -0.2001
#> 
#> (3,.,.) = 
#>  -1.0859 -2.3755  0.0482 -3.3380  0.1960
#>  -1.0397 -0.0820 -0.3077  0.3193 -0.6921
#>   4.5906  2.1765  0.3312 -1.6618  1.7852
#> 
#> (4,.,.) = 
#>   0.2182 -1.5933 -0.5580 -0.7248 -0.5321
#>  -1.0787  3.3994 -0.3214  3.0813  1.1568
#>   0.8071  0.4499  2.3006 -2.8498 -1.7655
#> 
#> (5,.,.) = 
#>  -1.4449 -0.6329 -0.1287 -0.3982  1.0380
#>   1.0912 -0.5622 -0.9204  0.3244 -0.3044
#>  -1.0486  0.8165  0.1689 -0.3840  0.4813
#> 
#> (6,.,.) = 
#>  -2.6542  1.6704  2.2287 -2.3475  0.7408
#>  -0.3354  1.5402  0.7258  1.0193 -1.2666
#>  -2.8163 -1.5496  2.5634 -4.7717  6.6664
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]