Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.0944 -0.2909 -0.2105 -1.6426 -0.0968
#>  -0.5095  1.5937  0.8975  0.2814 -1.4409
#>   0.5150 -5.5213  0.2000 -1.7644 -2.2119
#> 
#> (2,.,.) = 
#>  -1.7382  0.6224  2.3265 -2.3429 -1.5743
#>  -1.3328  6.3333 -2.5388  3.9005  1.1979
#>  -0.7319 -0.4659  3.1386 -1.7026  0.3340
#> 
#> (3,.,.) = 
#>  -0.4074 -3.2041  2.9957  1.5690  1.8030
#>  -2.4371  3.8698 -1.5384 -0.8017 -0.3044
#>  -0.2288  2.5465 -1.7237 -0.7973  0.0932
#> 
#> (4,.,.) = 
#>   4.5095 -5.7548  6.5482 -6.3678  4.3594
#>  -0.3411  0.7679  1.9262 -0.7732  1.8266
#>   1.1018 -2.3163  1.3986 -1.9219 -0.3864
#> 
#> (5,.,.) = 
#>  -1.0184  1.1959 -0.0646 -1.4272  2.3748
#>  -1.5536  2.2430  0.3510 -1.1306  0.5860
#>   0.5803  0.1126 -1.2039  1.1237  1.1808
#> 
#> (6,.,.) = 
#>  -0.2844  0.1368 -1.9277 -1.3138 -1.0772
#>  -0.6134  0.4425  1.7596 -1.0925  3.4618
#>  -2.0353 -1.9366 -0.6357 -1.6771  1.1650
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]