Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.4746  0.5945 -0.0201 -0.1926 -0.6512
#>   0.9788 -1.0945  0.4594  0.7711  1.3363
#>   0.9479  0.1428 -0.8401  1.1196 -0.6237
#> 
#> (2,.,.) = 
#>  -0.6925 -0.8889 -0.6500 -0.1405  0.2251
#>  -0.2841  0.4151 -0.9169  0.1002 -1.6851
#>  -0.3804 -0.5974 -0.7351  0.3581 -0.3178
#> 
#> (3,.,.) = 
#>  -2.0906 -4.0371  2.2421 -3.3429  0.5067
#>  -1.3921 -2.0061  1.3759 -4.4835 -1.6800
#>  -0.0733  3.3514 -1.2040  0.6346 -1.5579
#> 
#> (4,.,.) = 
#>   1.3503 -0.3285  2.1844  1.8009  0.1960
#>   1.9960  2.0881 -2.3838 -3.4338  2.9182
#>   0.8812  1.3110 -0.6154 -2.4275  1.5986
#> 
#> (5,.,.) = 
#>   0.0972 -0.3608  0.0713 -0.9189  0.0010
#>  -2.8737 -0.1697  0.2317 -0.3864  0.0084
#>  -3.3795 -2.3415  1.5879 -1.4607  0.8272
#> 
#> (6,.,.) = 
#>   0.8464  5.3155 -4.5109 -1.0336  1.4219
#>   0.2493 -0.6461  0.2626 -1.0153 -0.6072
#>   0.1784 -2.3402 -0.4460 -1.1350 -2.0306
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]