Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -2.2143  1.7719  1.6188  3.0621 -8.1082
#>  -1.2620 -2.9388  3.4296  0.2437 -2.6628
#>   0.0502  0.3764 -0.5258 -0.4394 -3.5912
#> 
#> (2,.,.) = 
#>  -0.0037  1.3654  1.6715 -1.9748 -1.9290
#>  -1.9355 -1.0478 -0.6043  0.2236  2.1453
#>   0.4972 -1.1440 -2.0165  2.3035  1.6108
#> 
#> (3,.,.) = 
#>   0.1253  1.0249  1.6767 -0.5957 -1.3360
#>   2.0601 -2.8280  0.6756 -0.6224  0.6610
#>   1.4143 -1.9032 -0.6156 -0.0225  0.9084
#> 
#> (4,.,.) = 
#>   0.1809  0.9202 -0.7155 -1.4528  0.0931
#>  -0.2888 -1.0523  0.9052  4.9961  0.6157
#>   0.8053 -1.3000 -3.3472 -1.1673 -1.7652
#> 
#> (5,.,.) = 
#>  -2.2039 -0.4892 -4.1672  0.2629 -0.1584
#>  -0.1604  0.3618  1.5722 -0.4458 -0.5621
#>  -2.2692 -0.3513 -1.6378 -0.2571 -2.2051
#> 
#> (6,.,.) = 
#>  -0.5727 -2.2596  2.2415  1.0783  0.4681
#>   2.5649  1.3908  1.1155  0.9822  0.9414
#>  -2.4552 -2.1914 -0.8433  0.0471 -1.3850
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]