Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.6505 -0.7004 -0.7792  1.6599 -0.3019
#>  -1.6880  0.1304 -0.3313 -0.2266  0.7083
#>   0.2316 -1.6097 -1.7490  1.0049 -0.1199
#> 
#> (2,.,.) = 
#>  -1.7051 -2.3641 -1.7626  0.5216 -4.8545
#>   0.6286  0.1783  0.1218 -2.0875 -3.6240
#>   1.8303  1.3116  2.1071  0.0437  0.7322
#> 
#> (3,.,.) = 
#>   0.5024  0.1472 -1.1584  1.0283  1.9841
#>   0.3800  0.8026 -0.6600 -0.2303 -0.0500
#>  -0.3295 -0.1209  0.6718 -0.7066  0.1575
#> 
#> (4,.,.) = 
#>  -1.4301  2.8562 -1.8568  0.1232  2.2052
#>   0.3714 -1.7766  0.6844 -0.0524 -2.6093
#>   0.3331 -0.1774  2.3108  1.0948  0.4637
#> 
#> (5,.,.) = 
#>   -0.1623   1.4274  -2.7220  -0.5072  -2.2261
#>   -0.8216  -4.5238   2.5375   0.5172   3.3511
#>    3.5272  13.6135  -1.3521   0.1904  -5.2431
#> 
#> (6,.,.) = 
#>  -0.8167  0.2299 -1.9972  0.0945  0.4907
#>   0.0371  0.1385 -0.3956  0.1027  0.8754
#>  -0.2149 -1.2332  0.9775 -2.6064  0.6805
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]