Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   4.0839  0.3060  3.0674  0.8315 -1.8627
#>  -1.8328  0.1390 -1.2855 -0.8570  1.0650
#>   2.5207 -0.2669  2.7782 -1.4959 -1.6465
#> 
#> (2,.,.) = 
#>   4.9159  0.5776 -2.1916  0.9965  0.7342
#>  -1.4647  2.4986 -2.7879 -0.1982  0.7234
#>   1.5127 -0.9170  1.6072  0.5824 -0.4416
#> 
#> (3,.,.) = 
#>   1.3851 -0.4706  0.3791  1.5733  1.3854
#>  -2.4897  0.1029 -0.7109  0.8157 -0.8968
#>   2.0512  0.7640  1.0718  2.4557  2.2143
#> 
#> (4,.,.) = 
#>  -4.6373 -0.9579 -2.5869 -0.6759 -0.1079
#>  -1.8254  2.2768 -2.4007  1.5769 -1.7978
#>   0.8265  1.1517 -3.4998  4.4610  0.9305
#> 
#> (5,.,.) = 
#>   0.3301  1.5827  0.0638 -1.7257 -3.4697
#>   0.5318  2.7392  0.8441 -1.3834 -4.0939
#>   2.1430  8.7127  4.4188  4.1841 -3.7935
#> 
#> (6,.,.) = 
#>   1.6351  2.7846 -0.0527  0.0701 -2.6310
#>   1.4487  0.9373  0.4893  0.2926 -1.8141
#>  -0.3778 -1.2264  1.0881 -0.3661  0.3789
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]