Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.1402 -1.7356 -0.5192  0.9518  1.9335
#>  -2.0102 -0.7882  0.9169  1.4785 -0.8083
#>   2.5931  0.4739  1.3266 -0.6537  0.4340
#> 
#> (2,.,.) = 
#>  -3.2509 -2.5386 -3.6321  1.6672  2.4769
#>   0.2246  0.0488  0.5932 -0.0512  0.6555
#>  -3.7229 -1.0174  1.2846  1.5726  1.4275
#> 
#> (3,.,.) = 
#>   2.4169  0.3963  0.6160 -0.0973 -0.5776
#>   1.0358  0.6476  4.0550 -4.2225  4.7929
#>   0.9643  0.9956  3.4709 -1.4028  3.0251
#> 
#> (4,.,.) = 
#>   1.9320 -1.0556 -1.9958  1.8633  1.9228
#>   5.5205 -0.5762  2.8956  0.4820 -2.1663
#>   1.6190  0.6412  1.3537 -0.9705  0.6149
#> 
#> (5,.,.) = 
#>  -0.8439 -1.5328  2.5983 -0.7047  0.5173
#>  -0.2644  1.7016 -4.8964 -1.2056 -1.4970
#>  -4.2618 -1.2604 -2.0974 -2.1510 -1.4934
#> 
#> (6,.,.) = 
#>  -3.4664 -2.3502 -0.1164 -4.3537 -0.3595
#>   3.2757  1.8611  1.1564  2.3401  0.6068
#>  -2.7158  0.2573 -1.1554 -1.5482 -0.5508
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]