Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -2.8700 -0.9683  5.6646 -1.8596  0.6673
#>  -3.8831  1.2459  3.0566 -1.8798  0.2069
#>  -0.6915 -0.2601 -2.0570 -1.8709 -0.1117
#> 
#> (2,.,.) = 
#>  -1.3717  1.9820  0.1821  0.3303 -1.4993
#>   1.1866 -1.5338 -0.8658 -0.8119 -4.1409
#>   0.6436  0.5655 -0.7570  0.1740 -1.2553
#> 
#> (3,.,.) = 
#>   0.8655 -2.3251 -0.6171  2.8091 -1.8179
#>   1.8614  0.2974  0.3940 -0.0243  4.3517
#>  -2.1512  0.8118  1.4861 -0.8497 -2.4281
#> 
#> (4,.,.) = 
#>  -3.8989 -0.2313  2.8089 -2.3285 -1.0212
#>   6.1204 -0.7244  0.1445  3.9934 -0.6046
#>  -0.3209 -0.3278 -1.1074 -0.4945  0.9000
#> 
#> (5,.,.) = 
#>   0.0030  1.3343 -0.0418  1.9889  0.7092
#>  -0.8330  0.2500 -0.3063 -0.1760  0.5397
#>  -1.9093  0.8658 -1.2742 -3.5239  1.2036
#> 
#> (6,.,.) = 
#>   3.0974  0.5067  1.4538  1.7986 -1.7424
#>   1.5339  1.7262 -1.5436  2.5340 -1.9714
#>  -0.5381 -2.7500  1.3435 -2.1175  3.2952
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]