Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   1.0938  0.2648  1.4444  1.0778 -0.2493
#>   0.9603 -0.3313 -1.4982  2.0433  0.4887
#>   0.8195  0.5883 -0.3460 -0.1353  0.8240
#> 
#> (2,.,.) = 
#>  -1.1263  0.5197  3.9355  0.7615 -1.4370
#>  -2.2292 -0.5493  2.7428  1.1504 -0.8392
#>  -1.0167 -3.6589 -2.5783  0.7771  2.1158
#> 
#> (3,.,.) = 
#>  -2.0330  2.1368 -2.5351 -1.9192 -1.9428
#>   2.7384 -1.2719  0.8224  2.0023 -1.4672
#>  -0.5121  1.2045 -1.6862 -1.2110 -3.6746
#> 
#> (4,.,.) = 
#>   1.3333 -1.1934  0.2345  2.3620 -0.7686
#>   4.9574 -1.4599 -2.4985 -0.2289  0.1484
#>   0.0734 -1.6730  1.4074  4.8781 -0.8294
#> 
#> (5,.,.) = 
#>  -3.8494 -0.4581  2.3288 -1.0476 -1.9649
#>  -1.8542  3.9632  3.4951 -3.4197 -1.5751
#>   0.1039  0.6185 -0.5317  0.4266 -0.4066
#> 
#> (6,.,.) = 
#>   2.0364  0.2277 -0.7989 -2.1442  1.1552
#>  -1.0693  0.6859 -0.3697  0.7187 -0.5548
#>   0.7470 -1.1323 -1.1386 -0.3713  1.0800
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]