Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -1.9322  2.5559 -2.3401 -3.5578 -3.6253
#>  -3.7618  1.3782 -1.1619 -3.6615 -5.6365
#>   3.3288  0.1814  1.2423  2.3876 -0.6516
#> 
#> (2,.,.) = 
#>   0.6606  0.7930 -1.1481 -0.0061  0.0266
#>   0.2788  0.0645  0.5471  1.3642 -2.1460
#>  -1.0426 -1.6798  0.8788 -0.5209 -1.0694
#> 
#> (3,.,.) = 
#>  -5.4645 -1.4420  1.5801  1.3998 -0.6355
#>  -0.6653  1.1949 -2.6872 -2.0414  2.7132
#>  -0.1243  0.4423 -1.0607 -1.6457  0.3215
#> 
#> (4,.,.) = 
#>   2.5849  1.8129  8.1627  0.1255 -1.4470
#>  -0.3047  1.1025  3.0054  2.0741 -4.5287
#>  -1.9628 -0.6420 -1.7031  2.6877 -0.9761
#> 
#> (5,.,.) = 
#>  -1.2937  1.3447 -0.3611  0.7749  3.4290
#>  -0.3924 -3.9315  1.5042 -2.0230 -1.6529
#>  -1.0125 -1.9205 -0.5642 -0.5719  3.9490
#> 
#> (6,.,.) = 
#>   1.4368 -0.2433 -1.4569 -3.1209  0.4375
#>   2.5299  1.3585 -1.3940 -1.3677  1.7398
#>   0.1190  0.6050 -2.3744  1.1975  0.3050
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]