Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.0244 -1.5870 -0.6310 -0.6157 -3.9231
#>   1.8074  0.3064 -1.7018  2.1651 -0.7376
#>   0.4620 -0.4821  0.0613 -1.9325  0.0280
#> 
#> (2,.,.) = 
#>  -0.0792 -2.0877  0.0863  0.0826  0.4013
#>  -0.4293 -3.9900 -4.0530  1.6303  3.7070
#>  -0.5347 -3.9027 -2.8218  1.1497  2.8304
#> 
#> (3,.,.) = 
#>  -0.0605  0.5058 -0.2159  0.8481 -0.7045
#>   1.0962  0.7111 -0.7096  3.3226 -2.2585
#>   0.5916  1.7654 -0.7623  3.6859 -2.5709
#> 
#> (4,.,.) = 
#>   2.5224  0.8557  1.1376 -0.7113  2.6355
#>  -0.8846  1.3513  0.1701  0.8008 -0.1431
#>  -0.4804  0.8825  0.7499  0.3629 -0.9982
#> 
#> (5,.,.) = 
#>   0.1650 -2.1347 -0.5480  1.6138 -1.5443
#>   0.3557  1.5878  0.7129 -1.6302  0.2680
#>  -1.3151 -2.0094 -3.6490 -4.8983 -0.0767
#> 
#> (6,.,.) = 
#>   3.9755 -1.9056  2.0757  2.1009 -0.9446
#>  -0.4834  0.4578  1.9932  1.2435 -2.0694
#>  -0.2216 -2.8194 -0.8827 -0.7385  3.3707
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]