Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -1.3923  0.4484 -0.4262  4.7244  1.4960
#>   0.5613 -0.0661 -1.5926  1.4606  0.6494
#>   0.1667  0.6875 -0.9684  2.2542  0.8652
#> 
#> (2,.,.) = 
#>  -1.6980 -4.8263  2.1021 -0.2908  2.3410
#>  -0.6149 -1.3224  0.7323  0.2477 -0.3758
#>   1.1534  2.3478 -1.5683  0.7967 -0.5200
#> 
#> (3,.,.) = 
#>  -1.3242 -0.8881 -0.4637  1.8384  2.0415
#>  -0.2831 -0.3787 -0.2723  0.3550 -0.1299
#>   2.8008  1.8375 -1.1136 -0.7940  1.8697
#> 
#> (4,.,.) = 
#>  -0.2962 -0.2035 -2.2392  4.0843  2.0645
#>  -1.3210 -0.3352 -2.9752  0.7838 -4.3309
#>  -0.6473  0.1151 -1.1459  0.8470 -1.9340
#> 
#> (5,.,.) = 
#>   3.4730  3.6567 -4.5992 -0.2824  0.0059
#>   0.6342  0.7410 -1.1148 -0.4520 -0.8414
#>  -0.2312  0.4767  0.2240  1.6631  1.2227
#> 
#> (6,.,.) = 
#>   0.1114  1.8731  5.0846 -0.9746 -0.6655
#>   0.2448 -1.1414 -0.8333  0.1154 -1.0540
#>   0.0714  0.6485  1.4454 -0.3753  0.0086
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]