Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.1701  0.5022 -0.5493 -4.1034  0.2003
#>  -0.8752  0.0890 -0.5236  2.4215 -1.0337
#>   1.0955 -0.3229  1.4970  0.0273  3.0858
#> 
#> (2,.,.) = 
#>   0.0917  0.0519  0.1202 -0.4341  0.6458
#>   1.8827  4.2374  4.3906  6.6846  0.6844
#>   0.3094  0.4602 -1.9815 -1.8131  0.9585
#> 
#> (3,.,.) = 
#>   1.8596  0.6174  2.6577  1.1650  0.5583
#>  -2.4656  5.3834 -1.6288  4.1379 -0.6755
#>   1.8422  2.6931  1.3657  3.3067 -0.5298
#> 
#> (4,.,.) = 
#>   3.3978  1.3835  0.0710  0.2427  0.2863
#>   1.0520  2.8934  1.5602 -2.4517 -0.0159
#>   1.2974 -2.0537 -2.4122  3.3800  0.0778
#> 
#> (5,.,.) = 
#>   2.0227  1.8600 -2.6866  0.3668  2.6019
#>   1.7555  0.6900  0.8245  1.1984 -0.5345
#>  -2.4997  0.0075 -2.8328  2.9271 -4.2191
#> 
#> (6,.,.) = 
#>  -1.7011 -1.1565 -2.4507  0.9536  2.8053
#>  -1.7546 -1.7679 -3.4048  1.0709  0.7592
#>   0.0820  0.1197 -0.8174  0.1352 -1.6979
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]