Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.3645 -1.8433  0.1677  0.3360  0.3245
#>  -1.9234  0.6624 -3.6687  0.5079 -1.4655
#>  -0.0945 -1.6363 -0.3605  0.9098  0.3724
#> 
#> (2,.,.) = 
#>  -0.0223  0.8187 -0.3184 -0.7006 -2.2330
#>   1.7870 -0.5922  1.3473  0.8649  0.9932
#>  -1.4434  0.9159 -2.1400 -0.5583 -0.0554
#> 
#> (3,.,.) = 
#>   0.0541  2.4823 -0.5852  1.7944 -1.2778
#>   0.1747  0.5843  1.8026 -2.2165 -0.8718
#>   0.1913  5.8561  2.9330 -2.5961 -3.3821
#> 
#> (4,.,.) = 
#>  -0.2115 -2.2234 -1.1389 -2.5940 -3.8356
#>  -1.0857  0.4416  1.8092  2.6947  2.4388
#>   0.6375  3.3025 -0.5113 -0.1599 -0.2668
#> 
#> (5,.,.) = 
#>  -2.1459 -1.0169  0.6573 -0.6782  0.4978
#>  -0.8722 -2.3131  2.2082  0.1991 -1.6948
#>  -0.3183  0.5040  2.4720  1.5241 -0.3744
#> 
#> (6,.,.) = 
#>   0.3236  0.2534  0.2618  0.7815 -0.9764
#>  -0.6168 -0.8232  0.2399 -0.4550  0.6804
#>  -0.8295 -0.7015  0.5302 -0.7039  0.5093
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]