Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -4.2813  0.0720 -2.0026 -2.6928 -0.0832
#>  -0.4810 -0.7298  0.2892  1.0552 -0.1184
#>   0.4625 -1.8480  1.2724 -0.7938  0.1533
#> 
#> (2,.,.) = 
#>  -1.2624  0.9343 -1.2891 -1.5213  1.6085
#>   3.9671  1.7466  4.7796  0.1222 -3.7581
#>  -0.3958 -0.5763  0.2437  0.3511 -0.2841
#> 
#> (3,.,.) = 
#>  -1.4388  0.0409 -1.4768 -0.5153  0.2711
#>   1.8789  1.4557  1.6537  1.0473  2.6276
#>   2.8756 -0.7316  1.8867  0.9782  0.7213
#> 
#> (4,.,.) = 
#>   2.9613 -0.4555  1.1490 -0.6949 -0.2851
#>   2.9527  2.1512 -1.3854 -0.6802 -0.3109
#>  -1.1982  3.3380 -4.9682  1.5008  0.0201
#> 
#> (5,.,.) = 
#>  -0.2292 -1.2144 -1.6006 -0.1814  1.0095
#>   0.7725 -1.5561  2.3841  2.0236 -1.8675
#>  -0.8258  1.2202 -3.0346 -1.7177 -2.4763
#> 
#> (6,.,.) = 
#>   0.0307 -1.7672 -2.2373 -1.3873  1.5137
#>  -0.0400  3.6629  0.5656 -1.9928  2.2293
#>  -0.2633  3.1683 -1.2382 -1.9809 -0.1085
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]