Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   4.0969  2.7207 -3.6008  1.3268  1.3406
#>  -2.7280  0.6885  2.8299 -0.0157 -1.6985
#>   3.6690  0.6897  3.6253  0.1628 -3.1317
#> 
#> (2,.,.) = 
#>  -0.9013 -2.2862 -2.6375 -3.8137 -0.3126
#>   0.9910 -1.1898 -0.1023  1.9063  0.6657
#>   0.0306  1.8628 -0.9448  0.3753  1.2503
#> 
#> (3,.,.) = 
#>  -0.6490 -0.3634 -1.0242  3.2870 -0.8392
#>  -5.1923  1.8453 -0.2658 -5.6125  2.6635
#>  -0.3532  0.6194 -0.8589  0.7935  0.8154
#> 
#> (4,.,.) = 
#>   0.4617  1.2714 -0.9518  1.9489 -2.3289
#>   0.3657  0.9549  0.2483  0.6426 -2.1853
#>  -0.5600 -0.2476  2.0935 -1.8868  0.6581
#> 
#> (5,.,.) = 
#>  -1.8169  0.5584 -1.2510  0.6767 -1.1486
#>  -0.9329  2.7748 -0.7906  0.5346  0.8042
#>  -2.9578 -0.5197 -1.8578  0.8856 -2.6397
#> 
#> (6,.,.) = 
#>   2.9565 -0.0147  6.5889  1.0653  3.9979
#>   1.7295 -1.2160  0.9521 -0.5427  1.2593
#>   0.1156  0.2761 -1.2101 -0.8653  0.5954
#> 
#> (7,.,.) = 
#>  -1.5607  1.5261 -2.9721  2.1814 -0.7757
#>   0.5217  0.4209 -1.6929  1.0194 -0.0481
#>   2.7703 -2.1920  4.9931  1.2913  1.6965
#> 
#> (8,.,.) = 
#>  -0.0993 -0.4571 -2.6658  0.7604 -3.6553
#>   1.0420  0.3870 -2.0977  0.8252 -3.3579
#>  -1.4386 -1.0430  2.1371  3.7170 -1.2288
#> 
#> (9,.,.) = 
#>   3.6815  0.1923  1.2587  0.2348  2.9124
#>   0.7078 -1.4278  2.6347  5.9435  1.0880
#>  -2.3576  1.2368  0.1069 -3.4711 -2.6293
#> 
#> (10,.,.) = 
#>  -0.2205  3.6978 -0.8393 -2.0953  0.0131
#>   0.9718  2.2484  2.8060 -0.7790 -0.2376
#>   0.6077  3.6983 -0.1727 -1.9426  0.7366
#> [ CPUFloatType{10,3,5} ]