Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   0.1712 -3.6537  5.5091 -0.2936 -3.4318
#>  -1.0453 -0.3621  0.1843 -0.2401 -0.4634
#>   1.4789 -3.2567  4.3227  1.8824 -2.3889
#> 
#> (2,.,.) = 
#>  -1.4371  0.7775 -1.8519 -1.1663 -1.3453
#>   0.0921 -2.6666 -0.4211  0.7238 -2.1462
#>  -0.0075 -1.8046  2.8154  2.7938  2.1235
#> 
#> (3,.,.) = 
#>  -1.2683  1.0170 -1.3618  1.6867 -0.0696
#>   3.7177  0.0995  2.4875 -0.8433 -2.9264
#>  -0.4639 -2.2270  1.1395 -5.4252  0.1221
#> 
#> (4,.,.) = 
#>   0.6288 -2.6898  3.4636 -3.1639  0.1162
#>  -0.6737 -0.1145  4.0699  0.7414 -0.2961
#>  -1.2150  1.5307 -5.1235  2.9537 -0.1016
#> 
#> (5,.,.) = 
#>  -1.7681  0.2570 -0.1601 -0.0594  1.9939
#>   0.0089  2.1735 -3.5718  0.0664 -1.8433
#>  -1.6833  0.7953 -1.8252 -0.8224  0.3828
#> 
#> (6,.,.) = 
#>   0.1825  1.2109 -0.6366  3.5054 -0.5445
#>   1.1740  1.7387 -0.9377  4.8903 -0.6481
#>  -2.5975 -0.5737 -0.0971  1.8130 -2.0836
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]