Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.3160 -0.5085 -0.7327  0.8042  0.0631
#>   0.5701  2.1744  2.6389 -1.2993  1.9544
#>  -0.6101 -0.2278 -3.5549 -0.7709 -1.9801
#> 
#> (2,.,.) = 
#>   3.2481  1.8537  0.8390  1.4679  3.0302
#>   1.8967  2.3853 -0.0836  0.3419  0.2783
#>  -2.8085 -1.4792 -1.3168 -0.9553 -1.7766
#> 
#> (3,.,.) = 
#>  -0.3593 -0.3952  2.2737 -0.5447 -2.0918
#>  -0.0578 -2.2845 -2.7107 -6.4866  3.1760
#>   0.2123  2.4331 -0.7537  4.9073  0.0330
#> 
#> (4,.,.) = 
#>   1.0315 -2.0024  2.8667  3.7158 -0.6353
#>  -1.9466  6.0312 -0.2808 -1.2152  0.6873
#>   3.9307  2.7684  2.0692 -1.3523 -1.1285
#> 
#> (5,.,.) = 
#>  -2.3257 -0.4380 -2.8465 -0.8286  0.8329
#>   1.1105 -1.8655  2.6563 -0.7038  1.9414
#>   0.7131 -0.0376  0.9661 -0.4644  1.1927
#> 
#> (6,.,.) = 
#>  -1.3707  0.0222 -1.2476  3.3315  1.5072
#>  -1.2768  0.2310 -1.7941  0.8014  0.9073
#>   1.6569  1.7268  2.1851 -3.8812  1.2073
#> 
#> (7,.,.) = 
#>   1.3032  0.0608 -0.3341  0.4274 -0.1581
#>  -0.3337 -0.9602 -1.4404  0.2605 -1.2473
#>  -1.8324  1.2747  3.1599  2.4771  0.3322
#> 
#> (8,.,.) = 
#>  -5.3437  4.6375 -0.2776  2.2516 -0.3956
#>  -0.2245  0.0662  0.2426 -2.1786  0.6463
#>  -0.0019  0.9403  0.8532 -2.0729  0.7451
#> 
#> (9,.,.) = 
#>   2.3853 -1.1255  2.6266  0.4292  1.0721
#>  -2.9049  1.2459 -0.2821  2.7301  0.5662
#>  -1.0272 -0.0106  4.3216 -3.5454  1.4419
#> 
#> (10,.,.) = 
#>  -0.7592  0.3726  1.2186 -0.2969  0.0381
#>   2.5995  1.5538  0.6243 -1.1154  3.6177
#>   1.5580  1.7301  1.8859 -1.2928  3.3290
#> [ CPUFloatType{10,3,5} ]