Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   0.6419  1.3284  3.3102 -0.5291 -1.7800
#>   2.0743 -0.3564 -0.1706 -0.2386 -1.8133
#>  -0.4790  3.1528  8.3413 -0.7924 -3.5600
#> 
#> (2,.,.) = 
#>   0.1928  2.0841 -0.8161 -0.7259  2.3095
#>  -0.3170 -0.2065  1.6670  0.0238 -1.7591
#>   0.5413  0.9045  0.5918  1.8780 -1.2756
#> 
#> (3,.,.) = 
#>   1.5051 -2.8823  1.1620  1.3208 -0.7439
#>  -2.2155  3.7675  0.1151 -1.1322  0.8716
#>  -1.9083  6.1350 -0.1389 -2.8162 -0.7059
#> 
#> (4,.,.) = 
#>  -3.3037  1.1402  0.8302 -7.1440  7.2233
#>   0.2758 -0.2074 -1.2486  1.4532 -2.9930
#>  -0.2537  2.1303  0.2281 -1.7080  1.7838
#> 
#> (5,.,.) = 
#>   1.5692 -1.3373 -3.1641  3.4408 -0.8630
#>  -0.9431  1.5218  2.2924 -2.1617  0.8752
#>  -0.4756  0.1777 -0.2121 -1.8192 -0.1974
#> 
#> (6,.,.) = 
#>  -2.8631 -3.5014  0.9239  3.0749 -1.9572
#>   3.8102 -1.8741 -0.1506 -0.5428  1.1031
#>   1.9997  0.0640  0.1686 -1.6846  3.2314
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]