Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   0.3495  1.7737 -0.4400  0.7316 -3.0826
#>   0.0400  2.1647  0.1768  0.5448 -3.8046
#>  -0.8782 -0.2431  0.5489 -0.2782  2.1542
#> 
#> (2,.,.) = 
#>   0.4697  0.0296  0.4184  0.3302 -0.3755
#>   0.9806  0.6332  0.7132 -3.1550 -1.1067
#>   0.7456 -2.2988  0.9508 -0.7650  1.2078
#> 
#> (3,.,.) = 
#>   0.2496 -0.6992  0.6392 -0.4245 -2.2663
#>   1.9787 -1.0775  1.3557 -1.2709 -0.5551
#>   0.9132 -2.1149  1.2287 -0.8972 -1.1045
#> 
#> (4,.,.) = 
#>  -1.9525  0.3721  0.1201 -2.5765  0.9066
#>   0.3230  0.4271  2.2535 -0.8834 -2.3902
#>  -1.9880 -1.5157  0.1048 -2.3725  0.7255
#> 
#> (5,.,.) = 
#>  -0.4972  1.0057  0.9598 -0.3900  0.7399
#>   2.1255 -0.5605  0.3080 -0.6936 -0.7266
#>  -1.8950  0.9910 -1.7374 -0.4989  0.0096
#> 
#> (6,.,.) = 
#>  -0.6971  0.0710 -0.9968  0.1094  0.6626
#>  -1.0029  0.2300 -0.7986 -0.1839 -0.4596
#>   3.4156 -0.1961  1.9166  2.1032  2.4884
#> 
#> (7,.,.) = 
#>  -0.5316 -0.0203 -1.1806  0.3093  0.7193
#>  -1.3470  1.4635 -0.3360  0.5590  1.5709
#>   0.7271 -1.7288  0.6919 -1.6975  1.1285
#> 
#> (8,.,.) = 
#>  -0.7130 -0.5114  2.3165  2.8829  1.6191
#>   2.2740  0.9807  0.5483 -2.5795  0.7973
#>   0.1278 -0.0177  0.1501 -0.0992  0.0018
#> 
#> (9,.,.) = 
#>  -2.3452  4.9098 -0.3775 -1.4020 -0.9351
#>   0.9564 -11.8976  0.5221  2.7766  1.1247
#>   3.7445  3.6276  0.1874  0.0093  0.4996
#> 
#> (10,.,.) = 
#>   0.1333 -1.0485  2.3658  1.9639 -0.5783
#>   1.4163 -0.6241 -1.1937 -0.0401 -0.0197
#>  -3.7338  0.7897  5.7202  0.9193  0.0914
#> [ CPUFloatType{10,3,5} ]