Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.0744 -2.6082 -4.7116 -4.7264 -1.0050
#>   0.2882  0.9952  1.4902 -1.2519  0.8178
#>   2.7170  1.3264  1.2194  1.6105 -0.0203
#> 
#> (2,.,.) = 
#>   2.3737  3.8480 -1.1084  3.6364  0.1347
#>  -2.5225 -4.2508 -0.0060 -3.5009  0.9092
#>   0.2141  0.4836  0.1678  0.4257 -0.7582
#> 
#> (3,.,.) = 
#>   2.5992 -0.7573  0.0194  1.1652  0.6825
#>  -1.6613 -1.9534  0.4084  3.0892 -1.5935
#>   1.8065  1.0539  0.8738  1.0689  1.9590
#> 
#> (4,.,.) = 
#>  -2.1304 -0.4446 -0.6988  0.0180  0.9186
#>   0.4290  2.9585 -1.3126 -2.5840 -0.4765
#>   1.7713  0.4207 -3.3061 -1.5998 -0.6618
#> 
#> (5,.,.) = 
#>   0.3111 -1.5078  0.5916 -0.4618 -0.6606
#>   0.4438  0.3253 -1.4351 -3.0346  0.5349
#>   1.3493  0.2155  0.5303  0.9589  0.3146
#> 
#> (6,.,.) = 
#>   1.7241  0.0536  1.5221 -0.7249  1.6364
#>   0.4524 -1.4485  3.9262  0.7106  3.9677
#>  -0.9066 -1.9000  2.4638 -1.8179  1.6563
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]