Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.5900 -1.2954 -0.0715 -0.0654 -1.7867
#>  -6.9084 -3.1853  0.7323  2.2998  0.2101
#>  -3.7767 -2.0487 -1.8702  0.4897 -1.6548
#> 
#> (2,.,.) = 
#>   2.2266  0.2543 -0.5135  0.9897  0.1917
#>  -1.4030  1.2229 -0.4108  0.2362  1.2039
#>  -2.9179  1.0857  0.0533 -0.5698  1.8549
#> 
#> (3,.,.) = 
#>   0.1172 -3.3883  2.3676 -0.2082  0.7851
#>   1.4991 -1.6143  1.3078  0.6926 -0.3589
#>  -5.1678  3.9503 -2.5645 -0.4452  1.9092
#> 
#> (4,.,.) = 
#>   3.8194 -3.9712  2.6337 -0.1505  1.1303
#>   3.9885 -2.6600  1.8485  2.5394  1.8649
#>  -2.7997  0.6649  4.1925  0.4525 -2.4720
#> 
#> (5,.,.) = 
#>  -0.1902 -0.2895 -0.1742  0.1561  0.0704
#>  -1.9171 -0.3562 -2.1316 -0.9746  0.6448
#>  -2.5856 -0.2863 -2.2217 -3.7801  1.9156
#> 
#> (6,.,.) = 
#>   3.5895 -0.2341 -1.5768  1.2276 -1.3787
#>  -2.1909 -1.8620 -0.4006  3.2136  0.9368
#>  -0.1297  1.6938  0.6465 -1.2248  0.2258
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]