Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.6419 -1.7000  1.9593  0.2037 -0.0268
#>   2.2066  2.1123  1.0726 -3.4956  0.8849
#>  -0.7154 -0.0897  0.8916 -0.5433 -1.5300
#> 
#> (2,.,.) = 
#>   0.7567  2.3021  1.4379 -0.5825  1.8628
#>  -2.2819 -0.2478 -0.1667  1.3805 -1.2194
#>  -1.6867 -0.2407  0.4612  0.4559 -1.6326
#> 
#> (3,.,.) = 
#>   0.8424 -2.4800 -0.3814 -1.7084 -0.8304
#>  -2.4122 -0.6510  0.2992  0.4659 -0.2688
#>  -3.8077  2.8146  0.8039  1.9638  2.4517
#> 
#> (4,.,.) = 
#>   1.0323 -1.7848 -0.1735 -0.3361  0.0372
#>   2.6094 -4.9350  0.4573 -0.7046  0.2635
#>  -3.4155  0.9380 -2.8630  2.7660 -1.9023
#> 
#> (5,.,.) = 
#>   2.2917 -0.8297 -0.0571  1.4344 -1.0633
#>   3.6172 -0.4836 -0.9313 -1.2559 -0.6526
#>  -2.9536  1.0550  1.3553  0.3170  1.0762
#> 
#> (6,.,.) = 
#>   2.4873 -0.1803  1.0847 -0.0266 -1.4320
#>   1.2493 -0.4212 -0.4391 -0.0985 -1.7677
#>  -1.1424 -0.0326 -0.2204 -0.5235  0.0304
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]