Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.8531 -0.8596  4.1955  0.9373  0.7889
#>  -2.1855  1.1320  2.0193  0.3556  0.2110
#>   1.0720 -0.6275 -0.8931 -0.5419 -0.1826
#> 
#> (2,.,.) = 
#>   -0.3220   0.6469  -1.1791  -1.4058  -0.7941
#>   -0.3105  -0.3146  -0.2758  -2.1812  -0.6268
#>   -3.1990  11.0644  -3.1932   3.8726  -4.6784
#> 
#> (3,.,.) = 
#>   0.1508  3.3688 -3.2357 -4.2239 -1.0033
#>   3.6709  2.6232 -0.4601 -0.6734 -0.8456
#>  -2.4414  0.2246  1.4326  2.8908  1.6616
#> 
#> (4,.,.) = 
#>   0.4737 -3.3060  1.0846  2.9995 -0.0123
#>   0.5777 -0.6861  2.8958 -2.0303  3.0517
#>   2.0083 -1.8582  2.7764 -0.7991  1.0703
#> 
#> (5,.,.) = 
#>  -2.8653  0.2511  2.2614 -1.3229 -0.2579
#>  -1.6650 -0.6148  1.2755 -0.2696  0.1141
#>  -1.5482 -1.7416  0.0272 -0.1736 -0.8695
#> 
#> (6,.,.) = 
#>  -1.5938 -2.2244 -0.2168  0.2077  2.4728
#>  -1.5074 -2.2563  0.4208  1.7533  1.8727
#>   1.9322  3.0611  3.6783  5.7015  5.5594
#> 
#> (7,.,.) = 
#>  -2.5312  1.6854  0.0703  2.7630  0.1071
#>   0.4723 -2.5291 -1.3869  3.0693 -2.4103
#>   0.4907  0.0351  2.5082 -2.0046  1.7938
#> 
#> (8,.,.) = 
#>  -0.5203  0.2183  0.4449 -0.4121 -0.6595
#>   0.1968 -0.9951  0.5894 -1.3364  1.7757
#>  -0.8528 -0.3432  2.5538 -0.6336  2.6878
#> 
#> (9,.,.) = 
#>   0.4622  5.1378 -2.0594  0.5971 -1.2780
#>  -3.6946 -2.8478  1.9603  1.0622  0.2568
#>  -3.6957  3.9388 -2.2408 -1.1617 -0.1115
#> 
#> (10,.,.) = 
#>   0.3278  0.2835 -0.2417  0.7938 -1.6645
#>   1.5445 -0.0678 -3.1863 -0.2677  2.1899
#>   1.0239  2.3181 -2.2358  1.3640  1.8623
#> [ CPUFloatType{10,3,5} ]