Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -3.2024 -7.6148  4.3702 -5.6429 -6.9038
#>  -0.2987 -0.0360 -0.4430 -0.1393  0.0781
#>   0.8201  1.1804 -0.5851  0.8954  1.2955
#> 
#> (2,.,.) = 
#>   0.0493  2.2746  0.7029 -1.5971  0.0149
#>   5.2546  0.3370  2.0622 -1.0184  2.1009
#>  -0.1337 -1.2488  1.0810 -0.7645  1.1350
#> 
#> (3,.,.) = 
#>  -1.0127  0.4178  0.7531 -1.6211 -2.1611
#>  -2.7020  0.6854  2.2722 -2.0922 -4.0586
#>  -0.6311 -0.5324 -0.0557 -1.1392 -0.7654
#> 
#> (4,.,.) = 
#>  -0.5658 -1.5348 -0.3974 -0.8457  1.4081
#>   0.2195  0.1102 -0.3070  0.0880  0.4718
#>  -0.9447  1.7458 -1.6602 -2.0345 -0.1711
#> 
#> (5,.,.) = 
#>   0.6328  0.8718  0.2306 -0.4081 -1.0625
#>   0.4642  2.1067 -0.7163 -1.6023 -3.3727
#>  -0.0486 -1.2502 -0.5548 -0.2660 -0.1620
#> 
#> (6,.,.) = 
#>  -1.8214 -2.5332  4.0614 -0.5801 -4.3507
#>  -0.8726  0.6566 -1.1563 -0.1508  2.0802
#>  -5.5626 -2.1082  0.9831 -0.9216  1.0816
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]