Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -4.3143 -1.4288 -2.8901  1.2246 -3.5967
#>  -2.3629 -1.2760 -0.3128  0.4496 -3.2110
#>  -2.5165 -2.0418 -1.2111  0.5410  0.4704
#> 
#> (2,.,.) = 
#>   0.2005 -1.8831 -1.8693  0.3386  2.4065
#>  -0.4507 -0.2797  5.1214 -0.0222 -0.1440
#>   1.8861 -0.5248  1.3518 -0.7618  1.8359
#> 
#> (3,.,.) = 
#>   3.1500 -0.2562  0.9322  1.8939  2.5073
#>  -2.9765  0.5298 -1.0403  0.0294 -2.5522
#>   5.5362 -0.4312 -0.9434  2.5797  2.0918
#> 
#> (4,.,.) = 
#>  -5.0034 -1.6248  0.6015 -0.9774 -0.1283
#>   3.1205 -2.7958 -0.6850 -0.2517 -0.2819
#>   5.0405 -2.6549 -0.3713  0.1941  1.0752
#> 
#> (5,.,.) = 
#>   0.5260 -0.2956  2.9246 -0.0223  0.8087
#>  -1.0080  2.0703 -1.2185 -0.4892  1.1755
#>   0.7336  0.1320 -1.4105 -0.2512 -1.0605
#> 
#> (6,.,.) = 
#>  -1.7508 -2.1923 -0.5212  0.0364 -0.7499
#>   1.2269 -0.8350  0.5142 -0.0501 -0.0092
#>   1.8582  2.7864  1.6327 -0.7579  1.5624
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#>   1.0149 -5.7616 -2.7425  0.9728 -3.7417