Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.2111 -0.4071 -0.4177  0.4692  0.0180
#>  -0.1129  0.8635  0.1443 -0.0267  0.7390
#>   0.5927 -2.4919  1.0297  0.1944  2.0226
#> 
#> (2,.,.) = 
#>   2.1392  0.1068  3.4495  0.4717  1.1548
#>   1.3179  1.1409  2.3881 -0.8562 -0.0957
#>   0.0249  1.8486 -2.8909  3.5791 -2.2251
#> 
#> (3,.,.) = 
#>   1.0592 -2.8352 -0.6328  1.5869  0.2503
#>   0.3270  0.1384  0.2119 -2.8849  5.7133
#>   0.1730  2.3325 -3.4228  3.2439 -0.2565
#> 
#> (4,.,.) = 
#>  -0.4845  1.8080 -0.1749 -1.5393 -0.1113
#>  -0.0023  2.2400 -0.7797  0.7470 -3.2968
#>   1.9389  1.8480 -0.4398  1.8934 -0.9689
#> 
#> (5,.,.) = 
#>   0.6062  0.7433 -0.9973 -0.0283  1.1732
#>   0.2694  0.1870 -0.2338  0.1843 -0.7659
#>  -0.5142  1.6874 -1.1413  0.8108  0.0085
#> 
#> (6,.,.) = 
#>  -1.0684 -0.5113 -0.3197 -3.7743  1.5850
#>  -1.6152 -0.2437 -2.9184  1.8048  0.2309
#>   0.4206  0.1017 -0.9891  1.9144  0.1638
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]