Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -1.2021  3.5793  5.1422  3.6396  3.5887
#>  -0.9474 -1.7029 -0.8850 -1.7209 -2.8198
#>   1.8686 -1.5457 -0.7297 -1.2525  0.5989
#> 
#> (2,.,.) = 
#>   1.7888  0.5601  2.2876  0.5266  1.8691
#>   0.4332 -4.3603 -5.3574  1.5241  3.5344
#>  -0.4044 -0.4682  1.1955  0.3176  2.6555
#> 
#> (3,.,.) = 
#>   0.5245 -0.7074 -0.5478 -2.0511  3.7212
#>  -1.2946 -3.3080 -1.9165  0.1011  5.3264
#>  -1.4391  1.4678 -4.2108 -0.4799  5.3518
#> 
#> (4,.,.) = 
#>   0.9250 -0.3344 -0.3653 -0.0240 -0.5394
#>  -1.2798 -0.0389 -0.9295 -0.0945 -0.3671
#>   0.6326 -2.4296 -5.1281 -1.5453 -3.0014
#> 
#> (5,.,.) = 
#>   4.5391  0.8035  1.6271  2.2518 -2.2185
#>  -1.0365  0.2945 -0.2870 -0.2120  1.0663
#>  -0.1211  1.2102 -1.3400  1.3690 -0.3011
#> 
#> (6,.,.) = 
#>  -1.0840 -1.1004 -1.7261 -1.1676 -1.1846
#>  -2.1634 -1.5177 -1.1333  0.3652  0.7774
#>  -0.2501  0.0154 -1.0261 -0.2047  2.2100
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]