Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.0973  0.1940 -5.7787  0.2809  0.9483
#>  -0.1443  0.4444 -0.7884 -0.2688  0.3710
#>   0.6820 -2.4157  3.8371  1.4502 -2.0028
#> 
#> (2,.,.) = 
#>  -3.2906  0.0541  2.8753 -1.4904  0.4839
#>  -0.7217  0.5053  0.6466  0.2738  2.5666
#>  -2.0292 -0.1021 -0.3136 -0.7082 -2.0892
#> 
#> (3,.,.) = 
#>  -1.2291 -1.2276 -1.7823 -0.5107 -2.0841
#>  -1.5090 -1.4030 -2.3768 -1.4097 -2.2101
#>  -0.5246 -1.9070  0.4806 -2.7532  0.3057
#> 
#> (4,.,.) = 
#>  -1.3839 -3.5002  0.0859 -1.3069  0.6241
#>   0.8389  5.1675  2.5961  2.2685 -0.2827
#>  -4.5020  0.5770 -1.3054 -0.3778  0.6620
#> 
#> (5,.,.) = 
#>  -3.5456  2.4140 -2.1759 -2.6073 -0.9152
#>  -1.1850  1.2241 -0.8260 -0.0899  0.1962
#>  -1.1884  0.4315 -0.8165 -1.1802 -0.6042
#> 
#> (6,.,.) = 
#>  -2.2887  2.5153  4.3024  0.4153 -0.7047
#>   1.7233 -1.6573 -2.8138  0.4212  2.6910
#>  -0.0001  0.9301  0.0799 -0.3625 -1.2676
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]