Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   3.4312 -1.2666  1.2968  0.8305 -1.4879
#>   0.1714 -2.8301  0.8567 -0.3634 -3.3591
#>  -2.3543  1.6249 -1.2102 -1.0107  0.7389
#> 
#> (2,.,.) = 
#>  -5.1471  0.0775  1.8359 -1.9040  2.0421
#>   2.5493 -2.6281  1.7577 -9.1933 -1.0981
#>   4.2538 -2.4902  1.0340 -9.9763  2.7646
#> 
#> (3,.,.) = 
#>  -0.3732  1.1829 -0.0516 -1.7252 -2.1453
#>   0.0823  0.2936  0.5376 -0.8380 -0.7494
#>  -0.6493  1.6940  0.2514 -1.5568 -1.0496
#> 
#> (4,.,.) = 
#>  -0.6855 -0.0102 -0.7460 -0.7539 -2.0002
#>   4.5053  0.7425  2.6127 -0.6158  4.2808
#>   1.5407  1.2687 -3.4872 -1.5633 -4.2524
#> 
#> (5,.,.) = 
#>   2.4774 -4.9256 -0.7260  1.0818 -1.1447
#>   0.8333 -2.3657  0.4411 -0.4759 -0.0147
#>  -1.6721  2.6056  0.3876  0.0292  2.4526
#> 
#> (6,.,.) = 
#>  -0.4873  0.5734 -1.6038  0.9589 -2.6955
#>   2.7677  0.7905  1.0521  0.5053 -3.5374
#>  -1.5428  0.5769  0.8425 -0.3127 -3.7346
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]