Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -5.0246 -3.2587 -4.1848 -1.8071 -7.4244
#>  -0.3891 -2.0437 -0.7088 -0.5581  2.2862
#>   0.5376  2.9507  0.8399 -0.0240  0.1556
#> 
#> (2,.,.) = 
#>  -0.1566  0.2805 -1.7868 -2.1582  2.1628
#>  -3.4514 -1.3700 -0.7705  0.7780 -0.4869
#>   4.6498  1.4248  2.6241  1.2286 -0.8411
#> 
#> (3,.,.) = 
#>   3.1656  3.9982 -0.3565  0.1931  1.6090
#>   1.5851  4.4938  1.8548 -1.1250  0.5786
#>  -5.1226 -4.4527  0.5422 -1.6835 -2.7219
#> 
#> (4,.,.) = 
#>  -2.6145 -2.1561  0.9454 -0.6230 -0.8108
#>  -0.6182  1.9840 -5.3632 -2.9583  0.2439
#>   2.0580 -1.4262  2.8204  1.2494 -0.5602
#> 
#> (5,.,.) = 
#>   1.7241  0.4225  1.9482  0.0339 -2.5404
#>   1.3542 -1.1359 -1.1598  1.7864 -3.1107
#>   1.9553 -1.3771  2.2024 -1.6107 -0.0866
#> 
#> (6,.,.) = 
#>  -2.8018 -0.0597  1.5974 -0.1753  0.4715
#>   0.2335 -1.5529 -2.7067 -0.4910 -2.5819
#>  -3.0006  0.8169  2.8601  0.2688  1.1767
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]