Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.7808 -0.9706 -0.5574  0.7867 -0.4783
#>  -1.2644 -0.6078  1.2979  1.4305  0.0474
#>  -1.8517 -0.9929  2.0764  2.5102  0.3357
#> 
#> (2,.,.) = 
#>   0.5046  1.1213 -0.1423  0.3917  1.2354
#>  -1.8104  1.4885  2.4840 -0.3098  2.0685
#>  -1.2042 -1.1195  0.4062 -0.7720 -1.0872
#> 
#> (3,.,.) = 
#>   0.8574 -2.4962  0.6592  0.0824 -1.0270
#>   0.7737 -3.7039  0.5798  0.2190  0.6547
#>  -1.1943  4.5904 -0.7956  0.2484  0.6868
#> 
#> (4,.,.) = 
#>  -1.2304  0.6470 -0.7290  2.7618 -3.8338
#>   2.2802 -0.1447  1.0624 -3.0783  2.5474
#>   0.6475 -0.0434 -1.6709  0.9355 -0.5424
#> 
#> (5,.,.) = 
#>   0.3282  1.0774 -0.1399  0.4602 -2.2072
#>  -3.1357 -3.7371  3.9290 -4.4379  2.3416
#>   3.1658  0.7530 -1.6884  4.6027 -0.0691
#> 
#> (6,.,.) = 
#>   0.8117 -4.7848 -0.6776  1.9106  2.7906
#>   3.5087 -0.4162 -1.1527  1.6681 -0.7276
#>   0.1870 -1.4934  0.0569 -0.7598 -0.0020
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]