Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.8987  2.5497  3.7402 -2.3147 -4.1687
#>   0.5965 -2.2718  0.8914 -2.0791 -3.6148
#>   1.1052 -1.2908  1.1921 -3.2175  1.0382
#> 
#> (2,.,.) = 
#>  -3.1856 -0.0342  1.9105 -0.0635  1.5932
#>   1.1596 -0.2731 -0.4323  0.2059 -0.9846
#>  -3.3647  2.1627 -3.6482  1.0649  3.1931
#> 
#> (3,.,.) = 
#>   0.7217  0.5180 -1.0626 -0.8526  1.7644
#>  -0.2838 -0.6806  1.1750  6.2963  0.4166
#>  -0.8383 -0.2520  0.0804  0.5900 -1.0115
#> 
#> (4,.,.) = 
#>   5.0813  2.1665  1.5988 -0.9296  0.9173
#>  -0.3149 -0.5044 -0.9569  0.1048  0.0344
#>  -0.0225  0.8183  1.5938  0.0848  0.1807
#> 
#> (5,.,.) = 
#>  -0.8740 -2.0919 -0.3980 -0.5307 -1.5364
#>   0.4463  1.9660  0.3882 -0.3425 -2.4962
#>   2.7074  6.0908 -0.7770 -0.8153  1.1137
#> 
#> (6,.,.) = 
#>  -0.8838 -1.1704  0.5643 -4.2573 -2.1919
#>  -1.7004  1.0061 -0.0978 -1.4097 -1.0342
#>   0.2296  0.6412  0.7690  1.0529  0.9783
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]