Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.7280  0.9436 -0.1228  0.9816 -0.0914
#>   1.6586 -2.4789  2.8045 -2.6180  1.1069
#>  -2.9505  3.9302 -4.1773  3.5007 -1.7443
#> 
#> (2,.,.) = 
#>  -2.7070  3.8163  0.0584  1.3521 -2.6039
#>   0.5601 -3.0613  1.3551  1.0338  1.3279
#>  -1.8417  0.2888  0.6426  1.2334 -0.7071
#> 
#> (3,.,.) = 
#>  -0.2427  3.2691 -1.1937 -2.8396  2.3501
#>   2.7606  0.8496  1.1377 -0.4408 -4.3179
#>   0.4774 -0.3253 -0.0272 -0.2383 -0.7301
#> 
#> (4,.,.) = 
#>   1.0547  2.9551  1.3465  3.6160  1.2265
#>  -2.1685  0.0608 -0.3515  0.4362 -1.1635
#>   3.3829  2.2294  2.3535  1.1913 -0.3952
#> 
#> (5,.,.) = 
#>  -1.9575  0.6167 -0.5407  0.3691 -0.0114
#>  -3.0705 -1.3112 -2.2224 -0.5282 -0.2241
#>  -0.6172 -6.1905 -3.6819 -2.5040 -0.5497
#> 
#> (6,.,.) = 
#>   2.0316  2.7404  0.6043  0.7247  2.6550
#>  -1.4636  2.3087  2.6454  1.2781  2.1567
#>   1.9715 -0.5036 -1.6738 -1.6671  0.2592
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]