Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.5151  1.6007 -1.1230  0.7798  1.1360
#>   1.9218  1.4516 -2.5457  1.7180  2.4529
#>  -1.1550 -3.5024 -0.0199  1.3121  0.0777
#> 
#> (2,.,.) = 
#>   1.7137 -0.7909 -3.9425  7.2362 -0.8604
#>  -0.9656 -0.0921  2.0320 -2.5164  1.4388
#>  -0.4006 -0.3572  1.6946  3.1665  2.9983
#> 
#> (3,.,.) = 
#>  -0.0213  2.2736 -0.7494  1.8375  2.4167
#>   1.0099  0.0329 -1.3700  0.2733 -1.5540
#>  -0.4305  1.3188 -0.0872 -0.2633 -0.0145
#> 
#> (4,.,.) = 
#>   1.2408  3.4738 -2.4550 -1.2750  1.0256
#>  -1.7644 -2.6844 -2.3245  1.2221  0.3710
#>  -1.3607 -0.8237 -2.5073  1.5691  0.2019
#> 
#> (5,.,.) = 
#>  -0.9564  2.5215  0.6320  0.1771 -0.9401
#>   6.9919  0.5102 -3.1090  4.6481  3.3803
#>   1.5660 -0.2725 -1.1453  1.0859  1.2879
#> 
#> (6,.,.) = 
#>  -0.0211  0.1352  1.5383  0.3761  0.4691
#>  -0.7906 -0.1114 -5.3267 -1.4508 -3.1438
#>   1.3321 -0.0646  3.4015  1.0413  2.0873
#> 
#> (7,.,.) = 
#>  -1.0986  1.4772 -0.4913  0.0232 -0.7973
#>  -1.3199 -0.6571 -2.4314 -1.7296 -1.8410
#>   5.6168 -0.5017  0.0619  3.2543  3.5388
#> 
#> (8,.,.) = 
#>  -1.7169  0.0210  1.2610 -0.6434 -1.0115
#>   1.8341  1.7985 -0.4882  3.0626  3.5985
#>   0.8934  2.2048  0.2374  3.2721  2.8436
#> 
#> (9,.,.) = 
#>   0.1234  0.8506 -0.3748 -0.5287 -1.7992
#>   2.8878 -2.0904 -1.9127  2.0575  2.5144
#>  -1.8298 -0.8147  1.7153 -0.3986  0.1786
#> 
#> (10,.,.) = 
#>  -1.7231  0.4335 -0.7661 -0.3453  2.6239
#>  -7.3980 -6.5172 -1.0728  1.9038 -1.4962
#>  -2.9779 -0.6423 -0.3737 -0.3348  1.8779
#> [ CPUFloatType{10,3,5} ]