Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.1234 -3.1273  0.2008 -1.3558 -0.4732
#>   0.7423  6.1216 -2.8801  3.1399  2.4366
#>  -0.6949  0.9990  1.6811  0.1138  0.4895
#> 
#> (2,.,.) = 
#>   1.5574 -3.4601  1.2530  1.3283 -0.8958
#>   0.7719 -5.5872 -0.0975  2.2027 -2.1097
#>  -0.1546  0.9997  0.3177 -0.2637  0.5993
#> 
#> (3,.,.) = 
#>  -0.1994  1.5404  2.0486 -0.2521 -1.8653
#>  -0.7448  0.3142 -1.1101 -0.2497  1.1790
#>   0.0065 -1.1634 -1.9460  2.2672 -0.1517
#> 
#> (4,.,.) = 
#>   5.0736 -0.3452 -2.8255 -0.5697 -1.1055
#>   2.4322  0.0003 -0.1397  1.4718 -0.4107
#>   4.3058  0.5044 -5.7464 -2.8272 -1.6157
#> 
#> (5,.,.) = 
#>  -0.4139  1.3894  0.4013 -0.6060  2.0300
#>   0.8999 -0.2861 -0.6740 -0.7017  1.8980
#>   0.0763  2.5662  2.4332  5.3143  1.1841
#> 
#> (6,.,.) = 
#>  -2.7758 -0.6689 -1.1816 -0.2657  0.8423
#>   2.4915 -0.4829 -0.9588 -0.9130 -1.3823
#>  -1.0433  0.2400  0.4706  0.5922  0.6126
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]