Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.6093  0.8086  5.3570 -2.8073  2.0675
#>   2.5843  1.7271 -0.5576  3.0029 -1.8664
#>   1.6627  2.7090  3.9513  0.9856 -0.4620
#> 
#> (2,.,.) = 
#>  -2.2873 -0.2711 -0.0378  1.9288 -0.8707
#>   1.0178 -0.2463  0.4475 -0.0567 -2.8737
#>   2.9413  0.7546  0.4377 -2.8247  2.3874
#> 
#> (3,.,.) = 
#>   1.0288  1.7689 -0.9037  2.3232  1.1233
#>  -1.0896  0.4361 -1.6440  2.3032  1.3043
#>  -6.3689 -3.5082 -7.0500  1.0184 -0.6198
#> 
#> (4,.,.) = 
#>   0.6887  0.6581 -1.6299  0.1589  0.4309
#>   0.0054  2.6651 -0.5742  1.7115  1.2705
#>  -0.5631  1.5075  0.6041  0.7526  0.6563
#> 
#> (5,.,.) = 
#>  -1.5020  2.6070  2.1117 -0.3953 -0.8431
#>   4.5708 -4.6112 -4.1152 -3.9279 -3.6776
#>   2.3795 -2.9667 -2.1736 -1.9612 -1.1729
#> 
#> (6,.,.) = 
#>  -2.2531  5.0107  3.7949  0.1615  1.9100
#>  -1.3263  5.3086  4.1507 -1.0310  0.4338
#>   0.8089 -1.8392 -1.2566 -0.1649 -0.3958
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]