Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.8384 -0.1952  0.1774  1.6042  0.8972
#>   2.3531 -1.8309  5.1782 -1.6516  1.6476
#>  -0.5540 -1.6220 -0.3605 -0.0326 -0.0169
#> 
#> (2,.,.) = 
#>  -0.6281  2.3685 -1.1817 -0.2875  2.1734
#>  -0.2191  3.9655  1.3001 -2.3680  4.1127
#>  -0.6894  0.4284  1.1494 -2.5441 -1.2825
#> 
#> (3,.,.) = 
#>  -0.9093  0.9979 -2.0516  0.5068  0.3517
#>   0.4783  1.3560  2.6104 -0.1419 -3.9667
#>   0.2034 -2.1492 -0.6915  0.8593  1.7467
#> 
#> (4,.,.) = 
#>  -0.2008  0.2899 -0.2545 -0.2294  0.1247
#>  -0.0362  0.7214  0.0558 -0.0742  0.2478
#>   0.3420  1.9232  1.0016  0.4165 -0.0429
#> 
#> (5,.,.) = 
#>  -1.0589 -0.5726 -1.2308  0.0961 -0.8764
#>   1.0866  0.3779  0.4383 -0.4596  1.0363
#>  -2.0862  0.5849  0.8633  1.1972 -1.7498
#> 
#> (6,.,.) = 
#>   1.0411  0.1372  0.0024 -0.0556 -0.7169
#>  -0.6091 -0.4672 -0.8070 -0.5645 -1.1906
#>   1.2638  0.3697 -1.5016 -0.1160 -3.0312
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]