Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   0.6792 -5.1876 -1.2301 -3.3695  1.4671
#>   0.6744 -6.9008 -0.0192 -2.1304  2.7430
#>   0.4909 -3.6668 -0.3715  0.0876  2.7980
#> 
#> (2,.,.) = 
#>   0.4140  0.7215  1.2229  3.1104 -1.4737
#>   0.1050 -0.1166 -1.0594 -0.6130 -1.0263
#>   0.8279 -0.8643 -1.0963 -2.4597  1.1111
#> 
#> (3,.,.) = 
#>  -1.2491  1.3795  0.8019  1.2122 -1.0811
#>   3.3244 -2.5085  1.5414 -0.5291  0.6228
#>  -0.7113 -0.2722 -1.4521 -0.4953  0.7381
#> 
#> (4,.,.) = 
#>  -0.1382 -3.1260 -3.2335  0.8283 -0.5450
#>  -2.7189  2.4072  1.5230  2.2398  2.1823
#>  -1.1497  1.1643  0.2889  0.3298  1.0628
#> 
#> (5,.,.) = 
#>   2.9823  0.8583 -0.8072  1.3162 -1.2535
#>  -2.6643  1.4032 -1.1614 -0.5791  4.5192
#>  -1.1748 -2.5254  0.0754 -0.8695 -0.8880
#> 
#> (6,.,.) = 
#>   0.7693 -1.9329 -0.4721  1.7455  1.9262
#>   1.4067  6.3190 -2.3209 -1.7615  0.0478
#>   1.0608 -0.4778  2.4423  0.3754 -0.8928
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]