Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -1.1882 -3.0333 -5.0421 -4.6863  2.5440
#>   0.7798 -1.5743 -0.6145  0.2195  0.1072
#>   0.5755  0.4607 -2.1603  1.9878 -1.0880
#> 
#> (2,.,.) = 
#>  -2.0138  0.4366  1.5147 -1.0158  2.9520
#>  -2.0151 -1.1505  1.5455 -1.9629 -0.3002
#>   0.0459  2.2645  1.8566  1.9842  0.0611
#> 
#> (3,.,.) = 
#>  -1.0914 -4.5264  2.5290 -0.3744  3.8573
#>   0.4572  2.4390  4.3803  0.6025 -5.6369
#>  -0.2076  0.3281  0.2035 -0.4604 -0.3794
#> 
#> (4,.,.) = 
#>  -1.8321 -1.0036 -2.9223  3.8825  0.4870
#>  -0.5251  1.2234  0.6256  1.2319 -2.2864
#>   1.1077 -0.5788  0.5579 -0.1880  2.7619
#> 
#> (5,.,.) = 
#>   0.6049  1.0420 -1.7571 -3.3479 -2.3853
#>   0.8128 -0.5160 -0.7522 -0.2871  1.5006
#>   1.2559 -0.6766  0.3166 -0.4112 -2.8615
#> 
#> (6,.,.) = 
#>   0.3724  4.0553 -0.5142  1.7023  0.1927
#>   0.2985  2.9111  0.6639 -1.5738  0.1515
#>   0.7600  4.6250 -0.3087  1.5450  0.7683
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]