Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   1.8201 -1.7821  0.9218  4.4766  0.2001
#>   0.6882 -0.9063  0.5525 -0.0053 -0.4730
#>  -0.6409  3.1045 -0.9623 -4.2760  1.1306
#> 
#> (2,.,.) = 
#>   1.3449  0.7551  0.2632 -2.0005 -1.4228
#>   1.6749 -0.8029  1.2482 -0.7894  1.4251
#>  -3.1548  3.9411 -1.8317  5.2569 -3.0536
#> 
#> (3,.,.) = 
#>  -3.1526  1.2057  0.9739 -0.8996 -1.1273
#>   2.2675  0.2292  3.3866 -3.0165  1.7322
#>   2.3694 -1.3872  1.0088 -1.2835  1.7872
#> 
#> (4,.,.) = 
#>   0.4718  0.5785  0.1799  2.7548 -1.4895
#>  -2.6165 -1.3670 -1.2415  3.2766 -0.9317
#>   1.4330  0.7341  0.5808 -1.3037 -0.0812
#> 
#> (5,.,.) = 
#>  -2.6988  1.0818 -0.9221 -1.0641 -3.9024
#>   3.4175  2.5223  7.8013  3.2353 -2.4696
#>  -3.5761 -0.0795 -2.5282 -1.1259 -2.7240
#> 
#> (6,.,.) = 
#>   1.4146 -0.0332  0.6781  1.8887 -1.3368
#>  -4.0467  5.1004  0.2413 -2.5265  0.4580
#>   2.9581  2.4302  2.4746  5.4929 -4.4928
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]