Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   0.0147  1.8811  1.2332 -0.8385 -0.0101
#>  -1.7285 -0.3175 -1.7672  0.8637  0.8253
#>  -2.6648 -0.2844 -1.8942 -0.0278  0.3561
#> 
#> (2,.,.) = 
#>  -0.8799 -2.1699  1.4866  0.9210 -1.4182
#>   1.5218 -2.8492 -1.0400 -3.3810 -7.1358
#>   2.4295  1.9677 -0.3305 -2.3584 -1.3695
#> 
#> (3,.,.) = 
#>   1.5236 -0.6395 -3.0054 -0.5978  1.6786
#>   1.4630 -0.8697  1.1087  0.1402  0.2898
#>   3.1859 -1.0100 -0.9824  1.1781 -0.2564
#> 
#> (4,.,.) = 
#>   1.3500  7.3376  4.6685  1.3030 -3.7199
#>  -0.2292 -4.2973 -2.0890 -0.0538  2.1145
#>  -0.1293 -2.1007 -1.5110 -0.4704  0.8762
#> 
#> (5,.,.) = 
#>   0.2203  0.8974 -0.6331 -1.2180  2.8686
#>   0.1081 -2.5854 -2.4306  0.1509 -1.9010
#>  -0.8497  0.2351 -1.5710 -0.8642  1.5445
#> 
#> (6,.,.) = 
#>   2.7487  3.2878 -0.1001 -1.8873 -0.7748
#>  -1.6058  1.4457  0.5421  2.9462  0.7485
#>  -2.2968 -2.6566  1.6972  1.6344  1.7036
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]