Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   0.6170 -1.9598 -0.4437  4.5799 -3.4198
#>   1.0492 -0.6978 -0.6653  2.2389 -1.3581
#>   0.9077  1.3516 -1.8503  1.9883 -0.2728
#> 
#> (2,.,.) = 
#>   2.2934  1.1812 -0.1058 -0.3892  1.0087
#>   3.4093 -0.6620  0.7275 -0.2157  0.0365
#>   2.0178 -1.0952  0.5899 -0.1999  0.9711
#> 
#> (3,.,.) = 
#>  -0.0619  0.6827  0.3728 -1.1292 -0.2736
#>   0.2212 -5.6072 -5.1019 -6.6551 -1.9555
#>   0.3011 -0.2123  1.1171  1.4403 -1.0654
#> 
#> (4,.,.) = 
#>   1.4700 -4.5666  1.5127  1.3673 -2.5677
#>   0.4972  1.8274 -2.2207  2.0025  0.8853
#>   1.9717  0.7444  2.1060  4.6298 -1.6877
#> 
#> (5,.,.) = 
#>  -1.6895 -0.4600  0.5584  1.2448 -3.0781
#>  -0.2603  0.2244  0.1640  0.4686 -0.0320
#>  -0.1355 -0.3180 -2.2371 -2.5381 -0.9418
#> 
#> (6,.,.) = 
#>  -2.2559 -3.1111 -0.5438 -3.5034  0.7106
#>   0.5545 -2.1832 -2.3873  3.0275 -1.8199
#>   0.9720  0.5235 -1.5777  3.1715 -0.7156
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]