Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   2.3679 -1.8087  1.0975 -0.7101  2.1890
#>   0.0584 -3.0391 -2.6255  0.1164 -0.6569
#>  -0.6480  1.1712 -0.3091 -0.0563 -0.2224
#> 
#> (2,.,.) = 
#>  -1.3245 -1.3144 -1.1419 -3.1994 -2.3111
#>  -2.6258 -1.7209  3.4333  3.1046  0.8050
#>   0.2066 -1.4443 -0.8492 -1.1600 -1.2640
#> 
#> (3,.,.) = 
#>   1.3153  0.1246  0.6213 -0.4338  0.7761
#>  -0.5403 -2.0143 -0.1320  0.5722 -0.0393
#>  -0.0170 -5.4284  1.3278  1.0764  2.3224
#> 
#> (4,.,.) = 
#>   1.9035 -1.5036  1.2797 -0.7558 -2.5329
#>   2.1391 -2.4038 -1.1221 -1.3231  3.8519
#>  -1.3818 -0.4016  0.1974  0.0259 -0.3726
#> 
#> (5,.,.) = 
#>   0.7024  0.2461 -0.9710  0.1676 -0.8775
#>   0.6292 -3.1356  3.1526 -1.8727 -3.4130
#>  -0.7589 -0.8034 -2.9059  1.1567  0.2028
#> 
#> (6,.,.) = 
#>  -2.3679 -1.6368 -1.4542  2.1490  1.9290
#>   1.1405 -0.7025  0.9726 -1.6620 -1.6778
#>   0.6991 -1.2995 -0.5418  0.9218 -2.0932
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]