Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.8862 -0.9087  0.7385  1.2052  0.0056
#>   0.0228 -0.0721  2.3088 -0.4169 -1.4653
#>   0.0228  0.9312 -1.1196 -1.5794  0.1471
#> 
#> (2,.,.) = 
#>  -0.7442  0.0878  3.8840 -0.8406  1.1391
#>  -1.0436  1.9557  0.3639  0.2688 -0.7340
#>  -1.7162  1.1210  0.6970 -1.4346  0.9473
#> 
#> (3,.,.) = 
#>   0.4031 -0.1111 -0.7085  1.0720  1.4118
#>   1.0403  0.5413  4.9920 -0.6175 -0.5261
#>   0.0348  1.6898  1.0675 -1.5129  4.8504
#> 
#> (4,.,.) = 
#>  -0.8432  2.9902 -1.4339  1.4484 -1.7167
#>  -3.5825 -9.3834  2.1846 -1.9744  7.1460
#>   0.6889 -0.8319  0.5597 -0.2446 -1.3186
#> 
#> (5,.,.) = 
#>  -2.0556  2.2547  2.4435  3.2254 -0.8585
#>  -2.5210  0.9259  2.5783  1.7302  0.5458
#>   0.0859  0.1409 -0.3586 -0.4898  1.1218
#> 
#> (6,.,.) = 
#>  -4.6970 -1.8992  0.4273 -0.0757  0.7697
#>   3.1124  1.7920  0.1389  0.1715 -0.1368
#>   1.1079 -1.8232 -1.9744  1.1793 -2.7303
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]