Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   1.4195 -2.0566  1.1233  1.5663 -0.5618
#>   0.8930 -0.1673 -0.0956 -1.9281 -1.9109
#>   0.3313 -2.4397  1.0230 -1.2041 -0.3325
#> 
#> (2,.,.) = 
#>   2.3913  0.2117  1.1612  0.4344  1.7204
#>  -2.0785  0.8348 -2.1061  0.3873 -1.9317
#>   0.3936  1.3240  1.5881 -0.1017  2.5500
#> 
#> (3,.,.) = 
#>  -0.2798  0.3112  0.1580 -0.6681 -0.1391
#>   1.8390  2.6383 -2.0487 -1.9022  2.3020
#>   0.3242 -3.6653  1.1965  1.2977 -0.5829
#> 
#> (4,.,.) = 
#>   2.3076 -3.1450 -2.2541  0.1355  1.5223
#>   0.4200  0.4918  0.5288 -0.4178 -0.2184
#>  -1.3898 -2.5293 -2.3780  2.4417 -0.3627
#> 
#> (5,.,.) = 
#>   1.1218 -1.0102 -4.3765  0.0512 -1.8404
#>   1.1181 -0.4914  0.7037 -2.8790 -0.3550
#>  -1.1440  2.3112  4.3230  0.0521 -6.0696
#> 
#> (6,.,.) = 
#>   2.7278 -0.5319  0.7295  1.9218  1.9330
#>  -0.1965  1.6640  0.9995  1.4135 -0.1888
#>  -1.8936  1.2740  0.4777  1.4698  1.1727
#> 
#> (7,.,.) = 
#>  -0.0193  1.5117  0.9961 -1.5562  1.7656
#>   0.7435  0.5142 -0.3803  4.5697  2.6297
#>  -0.0274 -0.0647  0.3493 -2.0790 -1.1441
#> 
#> (8,.,.) = 
#>  -0.8024  0.9187 -1.6129  0.6332 -0.2230
#>   0.1725  0.1283  0.1697 -0.0240 -0.1051
#>  -0.4049 -0.3281 -0.0214 -0.2423 -0.3108
#> 
#> (9,.,.) = 
#>  -0.1238  1.0081 -4.0071  1.2886  1.9155
#>  -0.2227  1.0897 -1.3067  0.7432 -1.5291
#>  -0.1947  1.1667 -3.4434  1.0161  0.5858
#> 
#> (10,.,.) = 
#>   1.9830 -0.4945 -2.0044 -3.5423  1.0548
#>  -3.8387  0.4144 -1.9572 -1.8388  0.6451
#>  -3.6524  2.9886 -1.3014 -2.3371  0.5552
#> [ CPUFloatType{10,3,5} ]