Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   1.7122  0.7848 -0.1484  0.7936 -0.9104
#>   1.1618 -0.2307  0.7712  1.8135  0.2053
#>   0.3911  0.7297 -0.4024  0.7398 -0.3564
#> 
#> (2,.,.) = 
#>   2.7940  1.6842  1.0058 -1.2417 -4.8352
#>  -2.8968 -2.1852 -0.6260 -2.8643  3.5983
#>   2.2882 -2.9523 -1.1212  2.8917  1.6476
#> 
#> (3,.,.) = 
#>   2.3983  0.2679  0.5268  1.0920  1.6682
#>   1.8218  3.2836 -5.8493 -0.6456  3.2245
#>  -2.8509 -2.1950  1.6186 -2.3727 -3.2672
#> 
#> (4,.,.) = 
#>  -2.7476  0.0365 -2.5831  0.4774 -0.0645
#>   0.2529 -1.8697 -0.3300  0.3459 -0.6969
#>   1.2514  3.6092 -2.2299 -0.1942 -1.1396
#> 
#> (5,.,.) = 
#>  -1.0106 -0.9630  1.5360 -0.1241 -2.4665
#>   0.0321 -0.9769 -1.0111 -0.2499  2.9527
#>   3.9985 -0.1025 -0.9286  1.0511  1.0170
#> 
#> (6,.,.) = 
#>   0.1837  1.5160 -0.5508 -0.0832 -0.6349
#>   0.7065 -1.0520 -0.2397  1.4061  1.6901
#>  -0.3777 -2.3602  0.3122  0.5371  0.3492
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]