Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   2.2053  0.2835  0.5091  1.6025 -0.0284
#>   0.1544 -3.5254  1.9447 -6.6992 -2.3619
#>   0.6665 -1.4795  1.5475 -1.7546 -0.0137
#> 
#> (2,.,.) = 
#>  -1.3675 -0.7354 -1.4035 -0.0933  0.4166
#>   0.3483  2.5762 -0.9904 -1.6425  0.5982
#>   3.1824 -0.7077  4.8386  1.2814 -1.8900
#> 
#> (3,.,.) = 
#>   1.1657 -0.1951 -2.3064  0.8478 -2.1498
#>   1.7505  0.1223 -1.8634 -0.8196 -1.7611
#>  -0.5169 -0.8659  0.0816  0.5778  1.8713
#> 
#> (4,.,.) = 
#>  -2.4877 -0.9516 -1.8776 -2.4159 -2.8908
#>   2.7519  2.9273  0.8042  3.3630  1.1305
#>   2.0653  3.2787 -2.2350  2.5360 -2.1346
#> 
#> (5,.,.) = 
#>   0.0109  1.1154 -1.4149 -1.0547 -0.8333
#>  -1.1873 -2.0306 -0.2625 -3.1134  0.6892
#>  -0.2104 -2.4043 -0.2108 -3.2834  1.0784
#> 
#> (6,.,.) = 
#>  -0.9284 -0.4830 -0.1323 -0.5943 -0.8122
#>   2.0574 -1.5667  2.2268  2.3019  1.0482
#>   1.3953  2.0917 -2.2261 -0.3266 -0.7013
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]