Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.2271  0.0313 -1.8900 -1.6144 -2.8701
#>  -1.6315 -0.3990 -0.8560  2.5197  1.3021
#>  -4.3597 -3.7522 -2.5730  3.0285  1.2370
#> 
#> (2,.,.) = 
#>   2.3135 -1.3628  2.1617 -0.6953 -0.2497
#>   5.8618  1.3248  1.1672 -0.3751 -1.7460
#>  -2.3633  1.6113  0.7497  3.8212 -0.9693
#> 
#> (3,.,.) = 
#>   0.0684 -1.9018 -2.5574  0.0821 -0.7550
#>   1.4943  1.7872 -1.3027 -1.5723 -0.0998
#>  -1.5374  0.6644 -1.2360  0.2017 -1.1060
#> 
#> (4,.,.) = 
#>  -0.7768  1.7641 -2.5238  0.3515  0.6012
#>   1.3639  0.0966  0.8255  0.4861  0.7236
#>   2.2733  1.8897 -1.0504  0.5577  2.6197
#> 
#> (5,.,.) = 
#>  -2.8031  3.3216 -0.6705  0.4330 -1.8755
#>   1.8285 -2.0468 -0.4630  0.0123  0.2558
#>  -1.0268  1.5579  1.8424 -1.8618  1.3176
#> 
#> (6,.,.) = 
#>  -1.7226  0.6723  3.3216 -1.9063 -0.0508
#>  -0.3829  0.6716  0.7875  1.1775 -0.2025
#>  -1.1981 -0.7447  2.4957 -0.8383 -0.6437
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]