Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   4.7101  1.2719  1.2553 -1.8853 -1.0864
#>   2.4834  0.9094 -1.2402  1.1361 -1.1803
#>   3.1523  1.0687 -0.0883 -1.2969 -2.8504
#> 
#> (2,.,.) = 
#>  -0.8574  2.7700 -2.4213 -0.7642 -0.5478
#>   0.5348 -1.7877 -0.2510  0.6889  1.0793
#>  -0.8720 -2.0425  0.3504  0.2370 -1.0926
#> 
#> (3,.,.) = 
#>  -1.0773 -2.7929 -1.8812 -1.1311 -0.3457
#>  -4.0041 -1.8878  0.1238 -0.1811  2.0995
#>   3.1337  0.3895 -0.4637  1.8006 -4.2420
#> 
#> (4,.,.) = 
#>   2.8617  2.8683  4.6492  3.4233  2.7582
#>   2.5538  1.2732  4.5067 -0.0615 -0.6925
#>   1.3528  0.8462  0.9483 -2.7502 -2.6945
#> 
#> (5,.,.) = 
#>  -0.5374  0.9753 -1.5163  1.1165  4.5170
#>   0.3424  0.0067  1.0019  1.2782 -1.2151
#>   0.6167 -1.8804  0.9421 -2.2355 -5.0822
#> 
#> (6,.,.) = 
#>   1.6531  0.7906  0.3333  1.3987  0.2642
#>   3.2508  2.9889  2.0874 -1.7649  1.2378
#>  -4.3740 -2.3497 -1.6982 -1.0726 -2.4815
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]