Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   1.2736  0.9201 -0.1982  0.3090 -1.1155
#>  -0.4817  0.0625 -0.2446 -0.1607  0.8456
#>  -2.7870 -0.2312 -0.3571 -0.2394  4.2301
#> 
#> (2,.,.) = 
#>  -0.7820  1.0801 -1.8622  2.1258 -1.9537
#>   0.1255 -2.3097  0.4353  0.4609  0.4977
#>  -2.4695  0.9127  0.6762 -0.3123  0.8728
#> 
#> (3,.,.) = 
#>  -0.7845  0.4580 -1.3759  0.3852  0.1021
#>  -0.5861  0.8414  0.0232  0.2696 -0.4354
#>   0.1559  0.4675 -0.7276 -0.1673  0.8826
#> 
#> (4,.,.) = 
#>  -0.3096 -1.2258  0.1173 -1.6109 -0.8184
#>  -0.2185 -0.6414  0.1624 -0.9407 -0.3955
#>   0.3925  1.0517 -0.2328 -0.9208 -0.5094
#> 
#> (5,.,.) = 
#>   1.1594 -1.4163 -2.6194  0.6246 -0.5216
#>  -0.8941 -2.0827  0.0684 -0.6858 -2.6017
#>  -0.7408  0.9875  0.5723 -0.2390 -2.1786
#> 
#> (6,.,.) = 
#>   1.9887 -0.0025 -0.2235 -4.4024 -2.2315
#>  -1.8462  0.5497 -1.4665 -1.2568  0.8760
#>  -0.2271 -1.8856 -2.8036 -4.4951 -1.3825
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]