Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.1622  0.4870  2.0216 -0.7480  1.4687
#>   0.1476  0.4297  0.4556 -0.2551 -0.5484
#>  -1.8190 -2.0044  3.0515 -3.3273  1.8178
#> 
#> (2,.,.) = 
#>  -1.0757  0.4098 -0.0455 -1.3411  0.3455
#>  -1.1427  0.9605  1.1994 -2.8423  0.2371
#>  -1.3813 -1.4591  0.1694 -2.0570 -0.4616
#> 
#> (3,.,.) = 
#>  -0.3388 -0.5046  0.4471  0.1255  1.7074
#>  -1.3354 -5.8392 -3.5493  3.0028 -0.8332
#>  -0.6413  1.5987  2.9495  0.5991  2.2497
#> 
#> (4,.,.) = 
#>  -0.6990 -0.1250  0.1409  0.6954  0.2256
#>   8.8326  6.1054  5.5011  3.5037  4.3255
#>   1.3245  2.8220  2.6863 -0.6588  1.4810
#> 
#> (5,.,.) = 
#>   1.1931  0.6623  2.7329  0.4973 -0.0886
#>   1.6736  2.7869 -3.6342  2.6148 -0.0516
#>  -0.4981  0.5364 -1.3862 -0.0322 -0.0776
#> 
#> (6,.,.) = 
#>  -0.6822 -0.1446 -1.3158  0.5777 -0.3739
#>   5.7460  3.7451  1.4297 -0.1960 -1.4568
#>  -2.0920 -1.5628  0.8456 -1.7255  1.0683
#> 
#> (7,.,.) = 
#>  -1.9099 -3.7356  2.5708 -1.6768 -3.0413
#>  -0.3728 -0.9098  3.5538 -0.1481 -2.8629
#>   3.1855  0.0806 -1.3456  0.0108 -2.6432
#> 
#> (8,.,.) = 
#>  -4.3877  2.1955  4.4050  0.1523  0.7371
#>   3.9588  0.9505 -1.1635 -0.1034  1.0614
#>   4.1848  2.9615  0.6999  1.2461  2.6099
#> 
#> (9,.,.) = 
#>   0.1354  0.7438  3.0482 -3.9158  3.8110
#>   0.7887  0.7585 -2.6928  0.3625 -1.3718
#>  -0.8756 -1.0277  3.4910 -1.7175  2.4794
#> 
#> (10,.,.) = 
#>   0.9482  1.1572 -2.1641  2.0364 -3.5202
#>   3.4038  1.1816 -0.3790  0.4290 -2.9176
#>   4.0664  0.4561  0.1105 -0.1488 -2.0254
#> [ CPUFloatType{10,3,5} ]