Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   0.7316  2.4590  0.9415 -0.1039 -0.1145
#>  -2.1439  3.4385 -0.0755 -0.1672  0.1102
#>   0.4215  2.8887  0.6078 -0.4311 -0.7242
#> 
#> (2,.,.) = 
#>   0.7330  1.1785  3.4146  0.9256  0.1689
#>  -0.1601 -1.4194  0.1998  0.7595  1.5668
#>   0.4047 -0.0604  2.0394  1.1163  0.4439
#> 
#> (3,.,.) = 
#>  -2.2357 -0.1967 -1.5658 -1.0586  0.3069
#>  -0.5385 -1.9202  0.4790  1.1242  1.8534
#>  -3.4626 -0.3825 -4.7449 -0.7861  0.6452
#> 
#> (4,.,.) = 
#>  -0.3765  0.9522 -0.3980 -0.0286 -0.0490
#>  -0.5970  0.6958 -0.8578  2.0269  0.4311
#>  -0.2650 -0.1635  0.0595 -0.8126  0.2835
#> 
#> (5,.,.) = 
#>   2.5650 -1.6514 -2.3147  1.7601  1.5574
#>   0.3527  0.5576  0.1843 -0.2398 -1.2896
#>  -4.3068 -1.4345 -0.1315  0.2475  0.8862
#> 
#> (6,.,.) = 
#>  -0.0345  0.8311  0.3262  0.0277 -1.2827
#>  -1.1413  1.3653  0.7656  0.1571 -2.7698
#>   0.7118  0.6553 -0.7636 -1.3719 -0.7081
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]