Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.5520 -3.5841 -1.2196 -2.9144 -3.7122
#>   0.5119 -0.8041 -0.6629 -0.2633 -1.3880
#>   4.8971 -8.4634 -2.5890 -6.1537 -2.9420
#> 
#> (2,.,.) = 
#>  -1.0033 -0.3995 -1.0962  1.1508  2.0070
#>   0.8567  1.1288  3.5404  0.4987  3.7852
#>   0.4277  0.7093 -0.6525 -2.1615  2.4539
#> 
#> (3,.,.) = 
#>  -0.1065  1.8141  0.3526 -1.0362 -0.1004
#>   0.0607 -2.8509  1.7009  4.9557 -2.8190
#>   0.6314 -8.5719  2.3733  8.2810 -3.9138
#> 
#> (4,.,.) = 
#>  -4.5639 -0.6038 -0.0531 -2.9602 -1.0162
#>   0.5369  0.5208  2.4177 -0.5260  0.1666
#>   0.5006  1.4455  0.0270 -2.6382 -3.0099
#> 
#> (5,.,.) = 
#>   1.7988 -1.2691  0.4784  0.5472 -0.1662
#>   0.1536 -1.6039  0.4195  1.0192  0.9321
#>  -0.5130 -3.1269 -0.6964  0.9002  1.7779
#> 
#> (6,.,.) = 
#>   0.3724 -1.6764 -0.6080 -1.5894  2.4989
#>   3.8099  3.2820 -2.4640 -1.3513  4.2587
#>   0.9404  0.7426 -0.0916  0.0377 -0.2244
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]