Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.2728  0.3329  0.3148  0.2418 -0.4662
#>   0.5190 -2.7376 -0.6840  0.4741  3.2800
#>  -3.8378 -1.8533  2.3469 -1.4773 -2.1407
#> 
#> (2,.,.) = 
#>   0.4744 -2.0795  1.7293  0.7364  0.6592
#>   0.9967 -0.2969 -2.2275  0.0368  1.3270
#>  -0.6186  0.3218 -0.5151 -1.6389 -2.9698
#> 
#> (3,.,.) = 
#>   1.3588 -0.3417 -0.5500 -1.3803 -0.1276
#>   0.3352 -3.6196  0.0343 -7.0251 -1.0040
#>   2.6105 -2.9322 -0.2034 -4.2093 -1.5587
#> 
#> (4,.,.) = 
#>   0.4325  0.0497 -1.9054 -1.4504  1.1657
#>  -0.6684  1.7983 -3.0887 -2.3614 -1.4598
#>  -1.2282  1.0689  1.6219  1.8710 -0.5221
#> 
#> (5,.,.) = 
#>   0.7162 -1.4029 -0.1230  1.3721  0.3804
#>  -0.5090 -0.0665 -0.6045 -0.4034 -0.6196
#>   4.3933 -1.3307  1.6651 -1.1499 -0.7551
#> 
#> (6,.,.) = 
#>   4.1973  2.9564  1.2474  4.1209  2.5958
#>  -2.1248 -1.3535 -0.9090 -2.6434 -0.9870
#>   1.3782  2.1178 -1.1425  2.5260  3.2836
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]