Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -2.4687  3.6672 -4.7455  1.1813  2.6297
#>   0.8096 -1.7949  3.8059 -5.1491  0.7476
#>  -3.3037  4.1075 -5.6462  2.1448  2.4805
#> 
#> (2,.,.) = 
#>   2.2961  0.4373  3.2282 -1.2307  0.5049
#>  -0.6697  0.2339  2.1627  0.1838 -5.8263
#>  -3.7580 -0.1849 -2.9531  2.3208 -5.3533
#> 
#> (3,.,.) = 
#>   0.5396 -0.5129  0.1252 -0.2428 -1.0784
#>  -0.1934 -0.7776 -2.9812 -1.6859  1.0741
#>  -0.4552  1.9240 -2.6213  0.5646  3.3452
#> 
#> (4,.,.) = 
#>  -1.4806  1.3696 -1.5042 -0.4190 -1.9077
#>  -0.6542  0.1521 -1.2293  1.3103  1.3903
#>   1.0279 -1.1699 -1.0293  1.0770  3.4013
#> 
#> (5,.,.) = 
#>   0.7087  1.0415  1.7461  1.3902  0.3702
#>   1.0146  1.3288  0.4922  0.2391 -0.4455
#>  -0.4661 -0.7190  0.9992  1.2079  1.1333
#> 
#> (6,.,.) = 
#>   1.3075 -0.4633 -1.0932 -0.2349  0.8751
#>  -1.9205 -1.4416 -0.3422  2.8619 -0.9541
#>  -0.1109  0.3627  0.0351 -0.4957 -0.6532
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]