Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.3913 -0.2402 -1.0412  1.4925  1.3509
#>   0.4697  1.4662  0.9735 -1.3564  0.8480
#>  -2.4180  0.9858 -2.8184  2.1294  4.5806
#> 
#> (2,.,.) = 
#>  -1.3185  0.4777  0.3904  1.1440  2.0388
#>  -0.1672  0.1524  0.0176  1.2274  2.0896
#>  -4.0426 -1.0682  0.1108 -0.0006  0.4369
#> 
#> (3,.,.) = 
#>   2.1742  2.3949 -1.0682  1.5554  2.5233
#>   2.0639 -1.0687  3.1722  0.3955 -0.2923
#>   4.4061  0.9213 -2.0298  1.2106  0.0715
#> 
#> (4,.,.) = 
#>  -0.0724  1.9612 -1.1481  2.1695 -0.7551
#>   0.4563 -0.4034 -2.3246 -2.0219  0.9602
#>  -4.4827  3.9261 -2.6190  4.9491  0.2450
#> 
#> (5,.,.) = 
#>   1.1302 -2.7710 -2.6533  0.3498 -0.1702
#>   0.8885  0.2603 -0.6814  2.9177 -0.2672
#>  -2.1188 -0.0168  1.2071  0.5230  2.7158
#> 
#> (6,.,.) = 
#>   1.7361  1.4311  0.0447 -1.7851 -0.8287
#>   0.6234 -0.3959  1.0953 -1.1155 -1.1444
#>  -1.9167  3.7018 -1.0548 -2.0382 -2.4910
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]