Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.2155 -0.5032 -0.6063 -0.0380 -0.7807
#>   2.3711 -0.0354 -0.8783 -0.9817 -0.1100
#>   3.0575 -0.1231  0.1653 -0.3688 -0.4321
#> 
#> (2,.,.) = 
#>   0.2573 -0.2565  0.1687 -0.2719  0.0322
#>  -0.5231 -0.0349 -1.1761  0.1076 -0.8617
#>  -1.7626  0.8112 -1.3264  2.7156  0.2704
#> 
#> (3,.,.) = 
#>  -0.0780  2.2695 -0.0074  1.3369 -0.8504
#>  -1.3963  4.1554  2.4209  0.9898 -1.0633
#>  -0.0792 -3.0282 -3.3298 -1.0606 -1.7473
#> 
#> (4,.,.) = 
#>  -0.0797  0.9994  5.4399 -0.1458  0.7935
#>   0.8894  0.5482  4.9815  1.6906  2.0919
#>   0.1898  0.6513 -0.2289  0.7105 -0.0875
#> 
#> (5,.,.) = 
#>   1.0029 -0.3565  0.5901 -0.5256 -0.8868
#>   0.9695 -0.0175 -0.1430 -0.2403 -0.5484
#>   1.0563 -1.8511  2.3293 -0.6962 -0.7106
#> 
#> (6,.,.) = 
#>  -0.0796  1.7661  1.1921  0.6781 -0.4170
#>   1.7989  1.1788  0.6259  1.9440  0.2720
#>   4.2231  3.0757  1.4561  3.8512  0.8912
#> 
#> (7,.,.) = 
#>   2.7903  0.1610  3.2409 -0.7119 -2.5587
#>  -1.6623 -0.1691 -1.9797  0.3478  1.4013
#>   2.2484  0.4374 -1.2629  2.7753 -1.6636
#> 
#> (8,.,.) = 
#>   1.7588  0.5309 -2.0600  0.1374  2.0724
#>   1.7669  0.0461  1.2269 -0.0576  0.4873
#>   0.5413 -1.5646  3.5831 -1.4262  4.1136
#> 
#> (9,.,.) = 
#>   0.4785  1.3489  0.6587  1.4300 -0.5847
#>   0.3270  0.7424 -0.5394 -0.5978  1.3623
#>   0.1295 -1.3712 -1.6074 -3.0900  1.8524
#> 
#> (10,.,.) = 
#>  -0.5709  1.0809 -2.1710  0.9024 -3.6912
#>  -0.9383  1.9862 -0.8454  1.1516 -2.0787
#>  -0.3711 -0.1399  0.6359  1.7822  0.0985
#> [ CPUFloatType{10,3,5} ]