Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   0.1611  3.2873 -2.7979  0.5577 -0.6578
#>   0.4613 -0.3262  0.8856  1.8179  0.5950
#>   0.3713 -1.0938 -0.0862 -0.2558 -0.2166
#> 
#> (2,.,.) = 
#>   1.7443 -0.3017 -4.1491 -3.1967 -0.9234
#>   0.3326 -0.1091 -0.1212 -1.4089 -0.1950
#>  -0.5194  1.0499 -2.9961 -0.3035  0.2586
#> 
#> (3,.,.) = 
#>  -4.2654  3.1454  5.5731  0.0939 -0.2378
#>   2.7372 -1.7179 -4.6256  0.3196 -1.2640
#>  -2.6296  1.7937  3.1864  0.0494 -1.7055
#> 
#> (4,.,.) = 
#>   0.2021  1.0450 -1.9129 -0.0005 -0.9617
#>   1.6322 -2.6664  0.4473 -2.3600  1.0421
#>  -0.8634  2.2971 -1.7369  1.3270 -1.4233
#> 
#> (5,.,.) = 
#>   0.9213 -1.8910 -2.8053  0.2184 -0.4071
#>   0.1079  0.2815 -0.0232 -0.0372  0.7164
#>   0.4005 -0.3513 -0.0439 -0.3791 -0.6494
#> 
#> (6,.,.) = 
#>  -0.4160 -2.6689 -1.4779 -0.2370  3.2740
#>   1.1628  2.6820 -0.3655  2.7943 -3.7204
#>  -0.0693  0.9010 -0.0905 -1.5431 -1.8316
#> 
#> (7,.,.) = 
#>  -0.6875 -6.1682 -0.6918 -0.4328  1.8613
#>   1.3210  0.0232  0.2676  0.9049 -1.5467
#>  -0.5596  5.2267 -0.2623 -1.0450 -0.3493
#> 
#> (8,.,.) = 
#>  -0.2389  0.4121 -0.2401  0.1842  0.9670
#>   0.4465 -1.6942  2.7656  1.1619 -3.8290
#>   0.6562 -2.4667 -1.6473 -1.5007 -2.2954
#> 
#> (9,.,.) = 
#>  -0.1681 -1.9747  2.9924  1.4263 -1.4143
#>  -0.9687  1.3590 -2.0486 -0.9397 -0.0612
#>   0.7662 -0.1328  1.9483  0.1000  1.3884
#> 
#> (10,.,.) = 
#>  -0.9815 -0.6052 -0.7606  0.2180 -3.1664
#>   0.0858  1.0146  0.1378 -1.6044  0.3625
#>  -2.1684 -2.1501 -2.8241 -1.3085  0.4587
#> [ CPUFloatType{10,3,5} ]