Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   0.1610 -2.7037  2.1053 -2.6791 -1.0705
#>   0.0556  0.2957  3.7280 -4.2083 -0.8613
#>   0.0154  1.5224  0.6869 -7.0450 -4.0122
#> 
#> (2,.,.) = 
#>  -2.3584  0.7346 -2.1549  3.4780  2.1660
#>   0.2249  1.5858 -1.4224 -1.0490  2.8688
#>  -1.4110 -2.6389 -0.0615  3.0121 -1.0156
#> 
#> (3,.,.) = 
#>   0.6997 -0.0285  1.5249  1.5090 -2.0568
#>   1.0638  0.0838  1.1323  0.1607  0.0322
#>   4.0740  0.3740  3.6667  0.3835  1.9800
#> 
#> (4,.,.) = 
#>   1.0528  1.0211  3.1406  0.9001  0.0666
#>   0.0827 -0.1492 -0.9904 -1.1549  0.4427
#>  -1.2775 -1.8778 -1.1774 -0.1045  0.8553
#> 
#> (5,.,.) = 
#>  -1.2796  1.3860  2.9826  2.7147  0.6371
#>  -0.9706  1.6976  0.1187 -0.2199  0.6832
#>  -0.8251  2.1150 -1.5528 -2.2786  0.4001
#> 
#> (6,.,.) = 
#>  -2.2963 -0.4141  1.3660  0.8131 -0.0678
#>   2.8062  0.4317 -0.7799  1.1269  1.8044
#>   0.0941  0.1338 -0.4474 -2.8567 -1.4313
#> 
#> (7,.,.) = 
#>   1.8403 -0.8114 -1.5227 -1.1454 -1.3980
#>   1.4667 -1.1762 -1.3695  0.2306 -1.3077
#>  -1.7362  1.3334  2.0960  0.5037  1.5335
#> 
#> (8,.,.) = 
#>   0.8667  1.3327 -0.4307  0.2257  0.1032
#>  -0.2995 -0.5799  1.9994  1.7789  0.2470
#>   0.7311 -1.9649  4.0189 -2.5746  2.3361
#> 
#> (9,.,.) = 
#>   2.7330  2.8420 -2.2068  0.4016 -1.0053
#>   4.5752  0.2123 -2.0005 -0.8215 -0.7418
#>  -0.1324 -1.8509  0.3986  0.7285  0.0226
#> 
#> (10,.,.) = 
#>  -1.8956 -1.2362  0.2815 -1.0489  1.3775
#>  -0.0783  6.8973 -2.6717  0.6828  0.6219
#>  -4.2229 -1.9789 -2.8656 -0.3915  3.3125
#> [ CPUFloatType{10,3,5} ]