Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -2.8152  0.8333  1.0645  0.1056 -0.6829
#>   1.8629 -1.3331  1.5816  0.3844  0.5258
#>   1.5794 -1.7217  2.7192 -0.4569 -0.6113
#> 
#> (2,.,.) = 
#>   1.5591 -2.1067 -1.8234 -1.0080 -0.0702
#>   1.4719 -4.5028 -1.4965 -2.0907  0.3962
#>   0.5250 -2.2535 -0.0396 -1.9079  0.4731
#> 
#> (3,.,.) = 
#>   2.9071 -4.1146 -2.0846  2.0110 -0.4510
#>  -6.1578  1.4703  3.5986  1.7022  1.8637
#>  -0.8493  0.7738  2.5128  0.8692 -0.9475
#> 
#> (4,.,.) = 
#>   0.3828  0.5320 -0.8238 -0.3859  0.4858
#>  -0.4611 -1.5660 -5.7779 -0.2619  4.7796
#>  -0.7629 -1.1790  8.4694 -0.0530 -4.3724
#> 
#> (5,.,.) = 
#>   1.4222  3.4772  1.0943 -4.2917 -0.1609
#>   1.4490  1.0410 -3.4809 -2.2866  1.3345
#>  -0.9281 -2.9380 -1.1055  1.0045  0.9061
#> 
#> (6,.,.) = 
#>   1.5081 -0.8433  2.0050  1.0449  1.6563
#>   0.8371  0.5009 -1.1153 -0.3256 -0.8160
#>   1.0537 -0.7656  1.5925  0.7122 -0.1464
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]