Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   1.1784  0.2382 -0.3883  0.2204  0.3190
#>  -0.0830 -1.2157  1.6173 -0.2895 -1.3444
#>   0.6287  3.0321 -1.6606  1.1931 -2.9053
#> 
#> (2,.,.) = 
#>   0.8887  0.1102  2.0054  0.5899 -1.3436
#>   3.5103  0.1237 -3.3990  0.2435 -1.3784
#>  -2.9417 -0.1186  2.3791 -0.4128  1.4076
#> 
#> (3,.,.) = 
#>   0.4662 -0.3148  0.9718  0.8051  1.4704
#>  -3.3985 -1.3815  0.1577  3.4148  2.2955
#>   0.7088 -0.2849  0.7783  1.0104  0.5507
#> 
#> (4,.,.) = 
#>   3.4023 -0.8562  1.2799  0.3845  0.1745
#>  -1.3072  0.7658 -0.8168 -1.0038 -0.9509
#>   1.9185 -1.0873 -0.4415 -0.1479 -0.1343
#> 
#> (5,.,.) = 
#>  -1.3202  0.0640  0.0636  0.4794  0.3211
#>   5.9041 -3.2587 -2.0043 -0.1278  2.1513
#>   1.5984 -1.3602 -0.6069 -0.2471 -0.4581
#> 
#> (6,.,.) = 
#>   0.7915 -0.5478  0.3761 -0.0903  0.0140
#>   3.0551  1.8199 -2.6535  0.3174  2.4060
#>  -0.9661  2.9711 -3.0543  0.0323  0.6633
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]