Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.0174 -0.8442  0.9072  0.3608  1.6278
#>  -1.1951 -0.3845 -1.1986 -2.4361 -3.5420
#>   0.5663  0.9647 -1.8817  2.3951  0.3643
#> 
#> (2,.,.) = 
#>  -0.4788  0.1179 -1.2891 -1.8978 -0.1956
#>   6.2415  0.6999 -0.5543 -2.6717 -2.7038
#>  -1.0921 -0.0635 -0.2984 -0.0839 -1.2453
#> 
#> (3,.,.) = 
#>  -0.9744 -0.2968  0.4403  0.0812 -0.7060
#>  -1.6348 -0.4177 -3.7320  3.4097  0.3371
#>   0.8926 -0.0233 -0.2649  0.0551  0.9324
#> 
#> (4,.,.) = 
#>   0.9380  1.0112 -0.2213  4.9864 -1.2364
#>  -0.9821  2.9615  0.7999  0.1353  0.0342
#>  -0.2093  1.1873  1.7727 -0.8469 -0.8161
#> 
#> (5,.,.) = 
#>  -1.3760  0.2310 -3.9775  1.8465  4.0140
#>   0.0049  2.3621  1.3890 -2.3586 -4.8561
#>  -0.1976  0.3768  0.0785  0.5466 -0.0035
#> 
#> (6,.,.) = 
#>   0.3467  0.6238 -0.1965 -1.2698  1.1437
#>  -0.5125 -0.5036 -0.1366 -0.4100 -0.7433
#>   0.5823  0.2898  0.7967 -2.4535 -1.4104
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]