Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   1.0900 -1.9240 -0.2227 -3.1819  1.3322
#>   0.0022  0.6853 -0.7482  4.3237 -0.3291
#>  -3.0421  6.4548  4.4374 -1.9231 -1.7421
#> 
#> (2,.,.) = 
#>  -0.2882  0.0531  1.4701  0.2267 -0.1609
#>   0.2272 -0.2940  0.0547 -0.1308 -0.7174
#>   0.1362  0.0815  0.3953 -0.0925 -0.5203
#> 
#> (3,.,.) = 
#>  -0.5324  0.2016  0.1730 -0.9402  0.5663
#>  -1.3400 -0.6987  1.0404  0.8518  0.7267
#>  -0.6032 -0.1382  0.4970  0.0006  0.7722
#> 
#> (4,.,.) = 
#>  -0.0472  0.1014 -2.3177  1.5874 -0.3965
#>  -1.7727  3.8799  0.6693 -4.1740 -1.7538
#>   1.6463 -2.1839  1.0308  0.7969  0.8886
#> 
#> (5,.,.) = 
#>  -0.1594 -1.1902  0.4163 -0.7338 -0.0654
#>   0.8736 -1.8466  0.2900  0.0261  0.9367
#>   2.5853  1.4835  0.5496  3.3293  1.2514
#> 
#> (6,.,.) = 
#>  -0.3963  1.1923  1.7156 -4.7848  3.6891
#>  -3.6429  0.0377  4.6655  1.7774 -1.0543
#>   0.1956 -3.7086  1.2700  4.5498 -1.3012
#> 
#> (7,.,.) = 
#>   1.3449 -2.3094 -2.4127 -2.9715  0.5498
#>   1.1891 -5.4430  0.5261 -1.7775 -0.9203
#>   0.2993  0.4477 -0.5923 -0.0613  1.0006
#> 
#> (8,.,.) = 
#>  -3.7112 -1.5218 -2.1636 -3.5276  1.9201
#>   2.1204 -4.9632  0.2223 -4.0829  2.4587
#>   1.9478  0.9245  2.5575 -0.1260 -1.9194
#> 
#> (9,.,.) = 
#>  -1.3706 -2.1395 -3.8553  3.4865 -1.9606
#>   3.3988  1.9200  0.6475 -1.4380  1.2849
#>   1.4036  1.7866 -1.6399  1.5816  1.5274
#> 
#> (10,.,.) = 
#>  -0.8774 -1.6786  1.4418  2.0458  0.4096
#>   0.9711  0.1049 -0.4370 -1.4520 -1.3917
#>  -0.7845 -1.7254  1.2893  1.9236  0.5884
#> [ CPUFloatType{10,3,5} ]