Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.0714 -0.7145 -1.3603  0.1641  0.4151
#>  -0.0491 -2.5066 -1.0912 -0.4717  1.4505
#>   0.1683  1.1866 -4.6118 -2.0754 -3.9239
#> 
#> (2,.,.) = 
#>  -0.8156 -0.4249  0.7588 -0.4588 -2.6486
#>   2.1486 -0.2721  4.2445  0.1724 -2.2127
#>   0.3952  0.1078  0.5283 -2.3556 -1.4126
#> 
#> (3,.,.) = 
#>   2.4193  2.9872  1.5585 -1.7832 -0.5912
#>   3.0611  4.1383 -0.9349 -1.1407 -1.4802
#>   4.3933  2.5561  3.8341 -3.0231 -1.5154
#> 
#> (4,.,.) = 
#>  -0.7972 -0.6787  2.0906 -0.6588  0.0099
#>   2.0664 -0.1028 -0.0269  2.0729 -1.0993
#>  -1.3210  0.6219 -0.3752 -1.7278 -0.3723
#> 
#> (5,.,.) = 
#>   0.3034  2.0494 -2.0569  1.4820 -1.6926
#>   0.3733  0.3854  2.7304 -0.3706 -0.1897
#>  -0.7856  0.4576 -0.2361  1.2863 -1.3240
#> 
#> (6,.,.) = 
#>   1.1991  2.0624  2.4139  1.3882  1.1210
#>  -2.1934 -0.1874 -1.7031 -1.7825  0.6203
#>  -1.8007 -0.1693  1.6759 -4.8228  1.6088
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]