Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   0.3700 -3.0060 -1.0092 -1.6025  1.4728
#>   3.2617  0.9689 -0.3809  3.5265  0.4934
#>   0.0943  2.3653  0.8591  3.1682 -0.8615
#> 
#> (2,.,.) = 
#>  -3.9789  4.6424 -1.7649 -0.7635 -1.6396
#>  -2.5276 -0.7958  0.1599 -0.3317 -3.1696
#>   0.5328  3.7936 -2.0327  1.7518  1.5904
#> 
#> (3,.,.) = 
#>  -2.4278  1.9287  1.6550  0.5142 -0.4275
#>   0.6695 -2.6340 -0.9966  1.1728 -2.3738
#>   1.2196  0.2636 -0.6881  0.7017  0.5842
#> 
#> (4,.,.) = 
#>   0.0374 -0.0381  2.3746  1.5070  3.0984
#>   1.7577  3.8318 -0.0803  0.9844 -0.6392
#>   0.7832 -0.2203  0.2906 -2.6535 -0.4359
#> 
#> (5,.,.) = 
#>  -1.0076  3.6204  0.3104  2.1792 -0.5203
#>   1.3552 -1.1576 -1.5993 -1.3777 -2.7461
#>  -0.9847  0.6508  2.5603  3.3885  2.0127
#> 
#> (6,.,.) = 
#>  -0.0733  0.0022  0.2050  0.0409  0.0732
#>  -0.2378  0.9177  0.4988 -3.9068  0.0750
#>   0.6920  2.0360 -0.5832  0.5973  0.3780
#> 
#> (7,.,.) = 
#>  -0.2335  0.6445 -1.8836  1.1289 -0.4609
#>   2.7404 -3.3569 -1.1286  0.9858  0.1012
#>  -0.1665  2.6133 -0.4266 -0.9073 -1.0464
#> 
#> (8,.,.) = 
#>   5.3133  5.7399  2.8018 -1.0021 -2.4450
#>   0.7708  5.0526  3.7111 -0.7381  1.2737
#>   4.9528  3.1199 -2.5786  4.5512 -4.5596
#> 
#> (9,.,.) = 
#>  -0.5227 -0.2928 -1.8288 -0.8849  0.1942
#>   1.9817 -0.2405 -1.0551  0.4138  1.0243
#>  -1.0623 -0.2895 -0.1806 -0.8472 -0.6496
#> 
#> (10,.,.) = 
#>    2.1213  -6.9586   0.4920   4.7345  -0.7934
#>    0.2769  -2.1206   0.3670   1.2241   0.4446
#>   -3.0214  11.0890  -2.8675  -6.6566   2.0433
#> [ CPUFloatType{10,3,5} ]