Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   0.9653 -2.1841 -2.6376 -4.5523  3.1903
#>  -0.7425  0.4076  0.6656 -0.5796 -0.1626
#>   1.3508  3.5563  4.0706  1.9233  1.1306
#> 
#> (2,.,.) = 
#>  -0.9671 -3.0728  0.9514  1.0583 -0.1486
#>   2.7918 -6.3967 -0.6650 -1.3200 -3.9787
#>   0.1694  0.3190  0.7739  0.6407  0.9300
#> 
#> (3,.,.) = 
#>  -1.2434  0.0569  0.1554  1.0196 -0.4837
#>  -1.2752 -0.0965  2.4298  1.1681  3.9179
#>  -1.2359  1.1244 -0.8078 -1.1239  0.8805
#> 
#> (4,.,.) = 
#>   1.9735  0.4722 -1.7435  2.4455  0.3010
#>   4.4907 -1.4203 -3.2271  4.9142 -2.3338
#>   1.2960  0.2903 -1.1938  1.2676  0.6124
#> 
#> (5,.,.) = 
#>   0.8163  0.2090  0.3955 -0.5253  0.8915
#>   1.0138 -6.2175 -2.7170 -3.5137  7.1676
#>   0.6918  1.7938 -1.0401 -0.5074 -2.5917
#> 
#> (6,.,.) = 
#>  -1.1056 -4.1173 -1.3330  0.3975 -3.4895
#>  -0.0659 -1.1106  2.6120  0.0325  0.5727
#>   2.0514  3.5833  4.4292 -0.4831  2.9907
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]