Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   2.4555 -4.5725 -1.9378 -3.9023 -4.3861
#>   0.7082 -2.6986 -0.0277 -3.1830 -2.1826
#>  -0.6084 -0.9691  0.5038 -3.7998 -1.1943
#> 
#> (2,.,.) = 
#>  -2.9989  3.3449  1.6452 -1.7685 -1.1187
#>   1.1784 -1.0793  0.8241  1.6479 -0.0230
#>  -0.2369 -0.9179 -3.0446 -1.3994 -2.2950
#> 
#> (3,.,.) = 
#>   3.0032 -0.7368 -0.3143  1.4207  1.2364
#>   0.9525  2.1434 -0.4662  1.5030  0.3653
#>  -0.3195 -1.9180 -2.1150 -1.5935 -1.3209
#> 
#> (4,.,.) = 
#>  -0.0651 -1.2340  4.4236  1.1432  0.0226
#>   0.6096 -1.1864 -1.4788 -0.1203 -0.4665
#>   0.3831  0.0857  1.7467 -0.3998 -0.0399
#> 
#> (5,.,.) = 
#>  -0.5438  0.6447 -1.7704  0.3834 -1.4191
#>   0.9289 -0.1566  5.8237 -0.0327  1.6746
#>  -0.4153 -1.2239  3.8334  3.7286  0.4878
#> 
#> (6,.,.) = 
#>   1.9975 -1.1670  0.9954 -0.2576  0.0818
#>   0.2629 -5.5256 -0.6223 -0.4300  2.4317
#>   0.3149 -4.2182  0.4338 -1.4049  0.1071
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]