Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.4061 -2.3787  1.2548 -1.8077 -0.0287
#>   0.2022  3.2849 -1.6574  1.6735 -1.1087
#>   0.5868  0.2474  0.6407  1.8896 -0.0420
#> 
#> (2,.,.) = 
#>  -4.3419 -3.5590  2.0802  0.9057  2.6583
#>   1.6591  1.0509 -1.7889 -0.3502  0.5144
#>   0.1051  0.0783 -1.4716  0.1535 -0.6196
#> 
#> (3,.,.) = 
#>  -1.6714  0.2212  1.1381  0.1909 -2.2391
#>   1.9241  2.0247 -0.8381 -1.7358  0.5374
#>  -2.4862  0.0144  1.5430  0.9950 -2.2532
#> 
#> (4,.,.) = 
#>   4.5618 -1.1125  1.4401 -1.2879  4.4474
#>  -0.6108  0.8324 -0.8609  0.6841 -1.4102
#>   0.1676  0.3992  0.6560 -0.6247 -1.5135
#> 
#> (5,.,.) = 
#>   1.0419  2.0329 -0.0568 -1.8999 -0.0314
#>  -0.6369 -2.1685  0.9180  0.3406  0.1738
#>   0.0299  1.7028 -2.1681 -0.5212  0.3557
#> 
#> (6,.,.) = 
#>  -0.3841 -0.6296  0.7154  1.0075 -3.1075
#>   0.5810 -0.3323  2.1062 -0.3591  0.6104
#>   0.0560  0.1037 -0.7536 -0.0309  0.3053
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]