Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   1.2591  0.3001  0.1015  0.5027 -2.4304
#>   1.5444 -0.1407  0.0820  1.2467 -2.5442
#>   3.0569 -1.3140  0.1676  1.2855 -1.9180
#> 
#> (2,.,.) = 
#>  -2.1428  1.2463 -2.4592  0.7950 -0.7167
#>  -1.4690  3.9446 -2.4137 -1.0588 -0.3028
#>  -2.9940  0.8062 -2.5886  0.7071 -0.2413
#> 
#> (3,.,.) = 
#>   0.0180 -1.0038 -0.2446  2.2563 -0.1566
#>  -0.2979  3.5523  0.7384 -3.4418  1.0501
#>   0.0373  0.5438 -0.2337 -0.3947  0.7673
#> 
#> (4,.,.) = 
#>  -0.4056 -1.8233  0.8321  0.7291 -1.7397
#>  -0.8351  1.0263 -2.1007 -1.7848 -0.4803
#>  -0.7223  1.0134 -2.0395 -1.9067 -0.1238
#> 
#> (5,.,.) = 
#>   1.6703  0.7556  1.5154 -3.2547  3.2134
#>  -3.6272  3.7532  1.5474  1.9527  1.0756
#>  -1.3641  2.6246  0.3844 -1.0004  0.1654
#> 
#> (6,.,.) = 
#>  -3.4169 -2.1290  2.0853  0.7349  1.4545
#>   1.6539  0.6925 -1.1852  0.8584 -0.9431
#>  -0.7507 -0.2674 -0.1330 -1.9551  5.5693
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]