Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -2.4100  1.2673 -0.7975  0.8494  1.5371
#>  -2.3447  0.5837 -0.5405  1.1916  1.4654
#>  -1.5199  3.4267 -2.8398 -0.7113 -0.7743
#> 
#> (2,.,.) = 
#>   1.9110 -0.7336  0.9123 -0.3316 -1.9496
#>  -1.8473  2.6482  0.2184  0.1186  0.2634
#>   0.6707  0.8563  0.8534  0.8005  0.0521
#> 
#> (3,.,.) = 
#>   0.1632 -1.0447 -1.2145  0.0209 -0.5465
#>  -0.0196 -0.9606 -1.6300  0.3031 -1.3656
#>  -3.2183  1.4066 -3.8577 -1.9513 -1.5528
#> 
#> (4,.,.) = 
#>   3.0806 -0.3046  5.1672 -2.7815  2.0509
#>   2.6683 -1.0913  1.2111 -2.1746  2.7662
#>  -1.0660 -0.3756 -1.4183  0.7866  0.6590
#> 
#> (5,.,.) = 
#>  -0.5655  0.7760  2.7473 -1.2948  2.7975
#>  -2.5263 -1.9137 -2.2205  0.0276  0.5405
#>   2.0899  1.0334  0.0493  0.5350 -0.5361
#> 
#> (6,.,.) = 
#>   2.0518 -1.7863 -1.9595  1.5060 -2.6687
#>   0.7095  0.2935  2.2067 -8.0321 -3.2461
#>   3.5431 -0.1307 -0.0708 -2.6324 -3.9452
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]