Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.9155  2.9859  0.5430  1.5232 -1.4726
#>   1.3944  0.2261 -1.7170  0.8783 -1.9424
#>   2.3203 -1.2154 -1.1066 -0.0754 -2.8728
#> 
#> (2,.,.) = 
#>  -1.6850 -0.9263 -0.5296  0.2516 -1.5576
#>  -2.1916 -2.1788 -1.4840 -0.9234 -3.1540
#>   6.5710 -1.7101  1.7712  1.9509  2.2016
#> 
#> (3,.,.) = 
#>   2.4883 -2.5915  0.6221  0.4214 -0.0279
#>   0.6174 -0.6851 -0.4567 -0.1252  1.0621
#>   0.5889  1.5068 -1.5833 -1.2361  1.0595
#> 
#> (4,.,.) = 
#>  -0.1865 -0.2605  0.7726  0.4835 -0.4439
#>   2.0242 -0.5842 -0.6835  0.3323  1.8845
#>  -0.5812 -0.2007  0.5561  0.7530 -1.2698
#> 
#> (5,.,.) = 
#>   0.1604  2.1922  2.5678 -1.7041 -1.8592
#>   0.9379 -2.9463 -0.6285  3.1554  0.0194
#>  -2.6759  4.7634 -1.8374 -1.8594 -1.4802
#> 
#> (6,.,.) = 
#>   2.7207 -1.8509  0.4083  5.2777 -0.6564
#>  -2.6219  0.5889  0.5385 -1.2372  0.2206
#>  -3.7760  1.3489 -0.8316  1.5938  2.3553
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]