Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   1.1740  0.5212  0.7081  0.2576 -1.5586
#>   2.0768  0.8108  0.0899  3.0517 -1.3026
#>  -0.7434 -1.6944 -1.7258  2.3408  0.5752
#> 
#> (2,.,.) = 
#>   0.0933  2.2385  0.9282  0.1679 -0.2846
#>   3.6990  0.7579  1.3712 -2.2231 -2.3123
#>   0.2620 -1.9341 -2.0340  1.1918  0.3823
#> 
#> (3,.,.) = 
#>   3.2031  0.1496  0.9826  3.0836 -0.4535
#>   4.4574 -0.7327  1.6426  3.2541  0.0190
#>   0.4421 -4.6722 -1.4998  0.0766 -1.5183
#> 
#> (4,.,.) = 
#>  -1.8287 -1.1416 -0.1849 -2.2599  2.1633
#>  -0.8010 -0.5000 -0.4788 -0.4281 -0.6089
#>  -1.6152 -1.5156 -0.5093  1.6502  0.2034
#> 
#> (5,.,.) = 
#>  -1.9619  0.1924 -2.9174 -1.1177  1.0446
#>  -1.2963 -4.0508 -1.1098  4.3943  2.3902
#>  -3.6833  1.0159 -7.4076 -3.4492  0.9439
#> 
#> (6,.,.) = 
#>  -1.3978  3.0976  1.4290 -0.2908  1.4796
#>   0.9205 -0.2046 -0.8656 -3.4330 -0.9101
#>   2.5668 -0.9253  0.3144  1.9151  0.2124
#> 
#> (7,.,.) = 
#>   2.8908 -6.1691  3.3674 -0.3120 -1.4232
#>   0.0557  0.9877 -0.5043  0.8321  0.3002
#>   1.9912 -0.1686 -0.3908  1.0894 -0.6215
#> 
#> (8,.,.) = 
#>  -1.1697 -0.5611  0.5927 -0.1646 -0.9640
#>  -2.8322 -0.7740 -0.4638  0.8177 -2.6096
#>   1.2229 -1.7940  0.1300  2.8135 -2.8847
#> 
#> (9,.,.) = 
#>  -2.0053 -1.4780  3.9945 -1.2368  0.8144
#>   1.7116  2.0322  2.3226 -0.6136  0.7513
#>  -2.0915 -0.0152  1.6682 -0.3998  0.6172
#> 
#> (10,.,.) = 
#>  -0.0213 -1.1985  0.8391 -0.2350  4.2565
#>   0.0084  0.4545 -0.1636  0.5349  0.5837
#>   0.0061 -0.1321 -0.0173 -0.3940 -0.8590
#> [ CPUFloatType{10,3,5} ]