Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.6892 -2.9728  0.0241 -3.2027 -2.6007
#>   0.1084 -0.1776  2.8059  0.4505  1.2633
#>  -1.7096  2.2337 -0.9336 -1.9072  0.2905
#> 
#> (2,.,.) = 
#>  -0.7902 -0.1712  0.4662  1.8570  0.8113
#>  -4.8960 -1.0055 -0.2960  4.2903  1.2674
#>   1.8304  0.6129  0.2860 -4.1792 -1.7203
#> 
#> (3,.,.) = 
#>  -0.1307  0.6116  0.5397  1.5403  0.7904
#>  -0.6892  0.9818  3.6747  0.7293  2.6694
#>  -0.0023 -0.2491 -0.0250 -0.7357 -0.1885
#> 
#> (4,.,.) = 
#>   0.5412  4.4480 -1.3073  4.0826 -2.3393
#>   2.7309  0.3867  0.7374  3.7352 -2.1847
#>   1.7106 -0.0881 -0.7346  1.4894 -0.9080
#> 
#> (5,.,.) = 
#>   0.6901  1.0170  0.9276  0.0349  0.2401
#>   1.5473  1.9444  1.3801 -2.1421  0.2485
#>  -1.1128 -1.3549 -0.7982  1.9063 -0.1952
#> 
#> (6,.,.) = 
#>   1.1190  1.4625 -1.7697  0.3677 -0.9900
#>   1.8781  1.5795 -0.6385 -0.5551 -1.7416
#>  -1.3444 -1.6924 -1.8308  2.0168  2.8198
#> 
#> (7,.,.) = 
#>  -1.3370 -0.0692  1.0358  0.7264 -0.0341
#>  -0.8309  0.0100  0.4540 -0.8498  0.8769
#>   2.8499 -0.1804 -0.8894  1.7145 -0.3989
#> 
#> (8,.,.) = 
#>  -3.0893 -5.6977  1.6339  1.2205  1.5077
#>   2.3056  4.2619 -1.7233 -1.5717 -0.3399
#>   0.4893  1.7143 -1.3193 -0.7460  0.7443
#> 
#> (9,.,.) = 
#>   1.2548  1.7629 -0.3286 -1.3930 -0.0344
#>  -2.3786  0.5149  0.5044 -1.4943  1.4068
#>  -1.3231 -1.7499 -1.8694  3.3903 -1.1771
#> 
#> (10,.,.) = 
#>  -0.9892 -0.5110  1.1834  0.8137  1.5746
#>   2.3613 -5.3717 -1.6194  3.2242  4.0656
#>   0.0826  0.5324 -1.4282 -2.5973 -4.3086
#> [ CPUFloatType{10,3,5} ]