Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   0.1453  0.0877  1.6693  0.7884 -1.0966
#>  -0.7165 -1.2355 -1.5465 -0.2080  2.5402
#>   0.4740 -0.3409 -3.4031 -2.1786  1.7633
#> 
#> (2,.,.) = 
#>   0.6115 -1.2790  0.5519 -0.6233 -0.2023
#>  -3.5756  2.8719 -2.1017  0.3419  2.6401
#>   1.5126 -1.3914  0.0036  0.1748 -0.0411
#> 
#> (3,.,.) = 
#>   0.0301 -0.2203 -3.7663  1.1684  1.4525
#>   1.1022  2.1627  0.7958  0.1482 -0.6971
#>   2.0497  1.8069 -0.4448 -1.3587  0.3439
#> 
#> (4,.,.) = 
#>  -0.1643  0.2097 -2.4252 -1.5508 -0.3109
#>   0.2350 -3.4606  2.4482 -0.2433 -1.5702
#>  -2.2366 -0.3759 -3.1636  0.6068 -2.7127
#> 
#> (5,.,.) = 
#>   0.8365  1.5047 -0.6938 -1.7994 -3.2894
#>   0.2568  0.1396  1.6462 -0.5341 -1.8920
#>   1.2696 -4.4636  1.2310  1.4872 -0.1022
#> 
#> (6,.,.) = 
#>  -0.7450  0.0147 -0.4720  0.3886  0.4526
#>   0.4246 -2.1329  1.4265 -1.8262  0.0309
#>   1.2612 -1.0251  0.1710 -0.8701 -0.3182
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]