Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   2.6942 -3.0468  0.7567  0.2750 -1.3102
#>  -2.3786  3.4788 -2.6529 -1.3693  1.2086
#>  -0.3782 -0.1453  3.7768  2.7126  0.7668
#> 
#> (2,.,.) = 
#>   1.6687 -1.2160  0.5969 -1.3080 -0.1282
#>   0.8045  6.0634  2.0593 -4.7690 -1.0410
#>  -1.7584 -2.9467 -0.6118  4.2310 -0.1426
#> 
#> (3,.,.) = 
#>  -2.9619 -6.5219 -2.3368 -4.9672 -2.1935
#>  -2.7847 -0.8775  2.2154 -1.0392  0.3848
#>   0.5238  1.5606  1.0221 -0.4063  0.4461
#> 
#> (4,.,.) = 
#>   2.8299 -1.3122 -0.6958  0.3799  2.3557
#>  -1.7404 -0.4543 -1.5766  3.9843 -3.8058
#>  -0.3454  0.6496  1.3749  0.1516  1.9568
#> 
#> (5,.,.) = 
#>   2.0962  2.8814  4.2240 -0.7003  2.9565
#>  -1.5433 -0.2140  1.2562 -0.0970 -0.2977
#>   2.2326  1.3056 -0.0499 -0.5778  2.1713
#> 
#> (6,.,.) = 
#>  -1.2630  0.5155 -1.7869  1.4281  1.2104
#>  -3.1393  1.2464 -0.4963  0.0178  3.0484
#>   0.1483  0.1381 -1.2193  1.1877 -0.0920
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]