Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -3.7991  0.3223  1.6515 -2.1851  2.3959
#>  -5.6617  1.3956  1.4255  1.4252 -0.8601
#>  -0.3430  0.2338  0.1384 -2.7117  3.0849
#> 
#> (2,.,.) = 
#>  -1.9203 -0.2002  1.0213 -0.1122  0.4836
#>   0.5572  4.0066  0.2350 -0.7961  2.6169
#>  -1.9068  3.7237 -5.7887 -4.3627  6.2690
#> 
#> (3,.,.) = 
#>   0.8620 -2.0236  0.8627 -0.6531  1.9907
#>   0.3180 -2.9061  2.4219 -2.1127  0.3423
#>  -4.7506 -2.2913 -1.7623  0.6794 -0.1887
#> 
#> (4,.,.) = 
#>  -0.1831  1.8780  0.5673 -2.2234 -0.1381
#>   2.9092  0.5791  0.5045  1.2369  0.2456
#>   0.9991 -1.6438 -1.7472  1.0120  1.7948
#> 
#> (5,.,.) = 
#>   0.7234 -2.3707  2.3557  4.7100 -2.0043
#>   0.1016  2.4971  0.9041 -2.0322  1.2531
#>  -1.0190  2.7782  0.6664 -1.3549 -0.3392
#> 
#> (6,.,.) = 
#>  -1.0665  1.0123  1.5556 -5.9237 -4.4342
#>   0.1291  1.0054  0.7914 -0.3597 -0.3959
#>   0.1573 -0.0306 -0.3887  1.8380  0.2064
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]