Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.1072 -0.0158 -0.3822 -1.1158 -1.5388
#>  -2.3003  1.0333 -0.2075  2.5923  2.6449
#>   0.8548 -0.3275 -0.1245 -0.3943 -0.0280
#> 
#> (2,.,.) = 
#>   0.2116  0.5509 -0.5643 -1.3638 -0.3587
#>  -0.8610 -0.1115 -2.4437 -3.2288 -2.7071
#>   1.4673  2.1193  0.2306 -0.0417  2.0109
#> 
#> (3,.,.) = 
#>   2.2168 -2.9271 -4.7758  2.2315  2.0748
#>   0.7208  1.9055  0.6278  0.4437  1.8757
#>  -3.8027 -2.2312  2.4808 -4.9469 -4.7457
#> 
#> (4,.,.) = 
#>   1.1149  0.2246 -0.3720 -6.4986 -0.3178
#>   1.3902  0.6695  0.1170 -3.3563  2.4197
#>   0.4576  0.1248 -0.3789  0.3077 -4.0639
#> 
#> (5,.,.) = 
#>   0.1249 -0.3852 -0.0780 -1.2264  0.7530
#>   0.3386 -2.3728  0.4904 -2.7538  1.1625
#>   0.0323  1.9691  2.0973  3.4143  0.0844
#> 
#> (6,.,.) = 
#>  -1.1584 -2.5816 -0.1011 -0.0523  0.9748
#>  -2.6882 -2.2901  3.1608  0.3695  1.4612
#>   1.3602  1.0784 -0.0143 -3.0652  0.8070
#> 
#> (7,.,.) = 
#>   2.7706  1.0168 -0.7959  1.7155 -2.9002
#>   0.3384 -0.3777 -1.3962 -5.9999  0.0783
#>   4.0748 -3.9354 -3.2110 -5.0182 -3.9793
#> 
#> (8,.,.) = 
#>  -1.9155  0.1193  0.2264 -0.1621  0.2574
#>   0.7682  1.4393  0.8583  0.3275  2.0760
#>   0.4676  0.2016  0.3376  0.0429 -0.2857
#> 
#> (9,.,.) = 
#>  -2.1372  1.0353  0.5086  2.9383  1.3306
#>  -0.7461  2.5929 -0.2001  3.9825  5.2761
#>   1.3688  3.0739 -0.9337  3.0454  7.1156
#> 
#> (10,.,.) = 
#>  -0.7785  1.1020 -0.5568 -1.8349 -4.0984
#>   1.0818 -0.9070 -2.1263 -1.2358 -2.8530
#>  -0.9459 -2.4846 -6.4171 -3.3630  4.7826
#> [ CPUFloatType{10,3,5} ]