Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.9590 -1.2515 -1.0249  3.2301  1.2805
#>   0.1300  0.6517 -0.0619 -2.7349 -1.7285
#>   0.6294 -1.4158 -4.1128  3.9076  4.2932
#> 
#> (2,.,.) = 
#>  -1.2928 -1.3528  0.9918 -3.6379  2.8243
#>  -1.4222 -2.6677  0.7633  8.3502 -1.4259
#>   0.9875  1.3014 -0.2067 -0.0467 -0.6837
#> 
#> (3,.,.) = 
#>   0.8767 -0.3727  1.7172 -0.8553 -1.7107
#>  -1.4083 -0.3836  1.0209 -1.1504 -0.0976
#>  -2.7342  4.3182 -1.4139  1.9295  1.2712
#> 
#> (4,.,.) = 
#>  -1.2600 -0.1751  1.5742 -1.1439  0.3023
#>  -1.4768 -0.3043 -1.3893  1.3801  2.4917
#>   0.9466 -0.3488  3.0798 -2.5673 -2.6139
#> 
#> (5,.,.) = 
#>  -6.3764  2.8008  1.4253 -3.6710 -1.9471
#>  -0.8083 -0.3745  0.8111 -0.0711 -0.8465
#>  -0.7483  1.3692 -1.1578  0.2362  0.8291
#> 
#> (6,.,.) = 
#>  -0.8950 -3.2627  1.9242 -1.0891  0.1099
#>  -1.9145 -3.1263  2.0976 -0.2527 -0.2650
#>   0.5701 -0.4545 -0.8419 -1.0382  1.5296
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]