Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.6715 -0.0222 -0.3052 -0.0132  0.3527
#>   0.0642  1.4605 -0.2209 -1.2005  0.5709
#>  -1.0371 -0.8810 -1.1245 -0.0821 -2.1417
#> 
#> (2,.,.) = 
#>   4.7513 -4.5550 -2.1236  5.3799  3.4089
#>   0.5636 -3.6664 -1.1034  1.4131  0.0695
#>  -1.2193  2.9154  1.1902 -5.7150  2.1859
#> 
#> (3,.,.) = 
#>  -2.5769  4.3691  5.2896  2.8948  1.6145
#>  -0.4768  1.2495  2.7310  0.7684  0.0762
#>   3.6303 -3.9340 -5.5709 -3.1720 -1.6945
#> 
#> (4,.,.) = 
#>   3.6557  1.9226 -2.5345  2.2762  2.9229
#>  -1.5318 -2.9690 -0.5611  0.6546  2.4618
#>   2.3266 -2.9180 -0.3267  1.6594  2.0594
#> 
#> (5,.,.) = 
#>   3.0726 -0.5080  1.5528 -0.8115  1.3036
#>  -2.4255 -0.0913 -1.9783  0.8748 -1.8288
#>   1.9180  1.2736  0.5709  0.4949 -0.0467
#> 
#> (6,.,.) = 
#>   0.6454  2.1866 -1.0247 -2.6391 -0.7656
#>  -1.3438 -2.0773 -0.1758  1.0906  0.6958
#>  -0.7500 -0.8464  0.6230  1.0011  1.2585
#> 
#> (7,.,.) = 
#>   2.7764 -2.0832  0.6146 -0.0110  1.1089
#>  -3.2453 -1.8295 -3.1091  0.0794 -0.3670
#>  -0.4551  0.8043 -2.0120 -0.3138 -0.0596
#> 
#> (8,.,.) = 
#>  -3.5918  1.9420 -1.9538 -1.1828  0.0897
#>   3.2484 -4.0688 -1.4353  1.3496 -3.0432
#>  -4.9755  7.0424  0.8682 -0.7064  4.2925
#> 
#> (9,.,.) = 
#>   0.3845 -3.2398 -2.1125 -0.3747 -1.6631
#>   0.8932 -3.2300 -1.0624 -0.8982 -0.7203
#>  -0.6003  0.1961 -0.2679  2.1828 -1.3433
#> 
#> (10,.,.) = 
#>   0.1576 -0.0329  4.1183 -1.2813 -0.6126
#>   5.1602  1.4532  6.5321  2.7236 -1.1874
#>   0.3057 -0.3436 -4.9753 -1.7966  1.4027
#> [ CPUFloatType{10,3,5} ]