Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -2.4351  0.6448 -0.3721  0.5038  3.3972
#>  -1.9048 -0.8543  1.6261 -1.5765 -0.4853
#>   1.9356 -0.3103  1.4963 -1.2757 -3.0478
#> 
#> (2,.,.) = 
#>   0.8280 -0.6412 -1.1642  0.2871  2.9873
#>   2.5545 -1.5857 -0.2575  0.4756  2.5070
#>  -0.4940 -0.9236 -2.3519  0.7609  0.8429
#> 
#> (3,.,.) = 
#>  -0.6334  0.9373  0.3465  0.2119  0.8987
#>  -1.5843  0.6361 -0.1507 -0.4122 -1.0236
#>   0.4071 -0.1572 -2.7423 -2.9364 -2.9148
#> 
#> (4,.,.) = 
#>  -0.4559 -2.0785 -0.0523  1.9143  1.1090
#>   2.9819  0.9378  0.5879  0.4181 -1.5213
#>   3.6420  1.3372 -0.4716 -0.1202  0.4701
#> 
#> (5,.,.) = 
#>   1.3283  4.6572  1.5107 -1.6146 -0.4999
#>  -0.4667  0.1770 -0.9330 -0.0658  0.9667
#>   0.7012 -1.3386  0.9346  0.3769 -1.4364
#> 
#> (6,.,.) = 
#>   3.5217 -0.0796  0.9754  0.6576 -1.8033
#>  -1.1112  0.4906 -0.5975 -0.1324  2.9746
#>   1.0962 -0.0462  2.6626 -1.1089 -2.9120
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]