Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   3.1442  2.4262 -2.2935 -1.3639 -2.1993
#>  -2.4462 -1.7138  1.8815  0.8984  3.3698
#>  -1.0652 -0.8479  1.6891  0.1200  4.3365
#> 
#> (2,.,.) = 
#>  -0.0271  0.2668  0.0764 -3.7463 -0.6649
#>   0.4892  0.5958 -1.6187  3.1196  0.9933
#>   0.2090 -0.5086  0.4513 -3.8265  0.0939
#> 
#> (3,.,.) = 
#>  -1.6169  1.5211  0.4808  3.8262 -2.2933
#>  -0.9707 -3.2139 -1.0269 -0.9894  1.3249
#>  -0.6936  4.9351  2.0400 -2.9567 -4.1846
#> 
#> (4,.,.) = 
#>  -0.3217  0.1556 -0.0135 -0.3032 -0.9760
#>  -0.4423  1.7172  0.0557  0.3300 -0.6263
#>   1.0265 -0.1551 -1.0761 -0.4859 -0.3006
#> 
#> (5,.,.) = 
#>   0.8418 -2.0823  3.9581  0.4504 -0.7363
#>   0.8824 -2.9564  6.6813  0.8545  0.4810
#>   0.1431 -0.9383 -0.1576  1.4777  3.4008
#> 
#> (6,.,.) = 
#>  -0.3246 -1.4146  2.1274 -0.5457 -0.6284
#>  -0.6453 -1.0518  1.1972 -0.5900 -0.8226
#>   0.4696  0.7103  0.1249 -0.2199 -0.5267
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]