Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -4.4960  0.2111 -3.3230 -2.0510  0.5564
#>   1.2464  0.1718  5.3455 -0.8726 -0.8918
#>  -0.2663  0.0291 -4.9619  2.7843  2.8515
#> 
#> (2,.,.) = 
#>   1.4827 -1.3098  0.5815  0.1660  0.1381
#>   3.6383  2.6728  2.7543 -0.3272  2.8014
#>  -0.5457 -0.6651 -0.5631 -1.3266  0.8276
#> 
#> (3,.,.) = 
#>   5.2296  0.1612  3.7191  1.2933  0.2849
#>  -3.8558  1.1754 -0.3137 -2.9254 -0.8890
#>   2.0237  3.3204 -5.8648  2.3751  3.8570
#> 
#> (4,.,.) = 
#>   0.6687 -0.6775  0.6857 -0.0664 -0.8321
#>   0.0314  0.1736  0.4087 -2.0943 -1.3628
#>   2.1523  0.5163  2.6655 -1.5949 -0.6409
#> 
#> (5,.,.) = 
#>   0.4935  0.3129  0.8493  0.7461 -1.4217
#>   0.4350 -0.5919  1.2704 -1.2751 -0.5305
#>  -1.4301  0.2547 -1.0649  3.1855 -0.1830
#> 
#> (6,.,.) = 
#>   0.3144 -0.5894 -0.0752  0.2062 -2.5121
#>  -1.6230 -4.8498  2.8722 -0.4579 -2.5725
#>  -1.8871 -3.4050  2.6670 -0.5423 -1.0757
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]