Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.6652 -0.3598 -2.1806 -4.1993  2.9668
#>  -1.0435 -0.9175  1.6755 -1.7232  1.9638
#>   3.6777  1.2233 -2.9693  3.2889 -2.4180
#> 
#> (2,.,.) = 
#>  -0.9944  0.0309  0.6044  1.4010  4.8376
#>   1.3878 -0.9107  0.2784 -0.2577 -0.2584
#>   2.4233 -0.5180  0.3729 -0.3207 -2.9159
#> 
#> (3,.,.) = 
#>  -1.0007 -0.0578 -0.1773  0.2771 -0.4825
#>   2.6303  0.5881 -0.3797 -0.4446 -2.0607
#>  -3.5291 -0.5542 -2.0220 -2.7498  0.3675
#> 
#> (4,.,.) = 
#>  -0.1144 -3.0947  0.0484 -1.8858  1.2206
#>   1.1849  4.3189 -0.1644  0.4036 -1.5877
#>  -0.1025 -1.7215 -0.0355 -0.2539  0.6511
#> 
#> (5,.,.) = 
#>  -5.1879 -0.2647 -6.0604 -6.9966 -0.0508
#>  -0.0096  0.5270 -1.4391 -1.3797 -0.3993
#>   0.5049 -0.2360 -0.8153 -3.9225  1.2679
#> 
#> (6,.,.) = 
#>  -0.2078  1.7032  1.5880 -2.4392 -0.5543
#>   2.0159 -0.7288  1.7270 -0.6159 -1.4762
#>  -3.4412  1.4488 -2.6631  0.1377  2.4075
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]