Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   0.3870  1.5759 -0.1566  0.8972  1.4431
#>  -0.1486 -0.7542  0.5905  0.0388 -0.4733
#>  -0.9050  0.9451 -0.1516  0.3590  0.8491
#> 
#> (2,.,.) = 
#>   1.2276 -2.9903  4.5347 -0.4500 -0.0620
#>   0.2882 -2.0143  2.2114  0.1295  1.1966
#>   0.2034 -1.1649  1.0773 -0.1614  0.6202
#> 
#> (3,.,.) = 
#>   0.5505 -1.2054 -1.6786  2.8877  0.0487
#>  -0.5210  0.5001  1.0324 -1.0671  1.7574
#>   0.8336 -0.8942 -0.1995 -0.2330 -1.2551
#> 
#> (4,.,.) = 
#>   0.9112  0.9859  0.1070  0.0307 -1.7537
#>  -2.2679 -1.4426  0.0132  0.5149  0.9917
#>   3.5713  1.4396 -1.6391  1.8077 -3.0790
#> 
#> (5,.,.) = 
#>  -1.7602 -0.7350 -1.3093 -0.1648 -0.2619
#>  -3.1693 -0.4249  1.7375 -1.7096  6.1813
#>   1.3297 -0.0092  1.4487 -0.4289  1.9521
#> 
#> (6,.,.) = 
#>  -2.2846  1.4348 -1.7628 -0.3691 -1.8725
#>   2.4353 -0.8051  0.6766 -0.7986  4.9902
#>   6.3195 -0.1649  2.9648 -1.1189  6.4374
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]