Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.1953  0.4156  2.2961 -0.4067 -1.4750
#>   1.1822 -1.3655 -2.7099 -1.8576 -2.0776
#>   1.4456 -0.6060 -2.1816 -1.2418 -0.1987
#> 
#> (2,.,.) = 
#>  -1.9740  2.4815 -4.5715 -1.9184 -1.1006
#>  -0.1496  2.7806 -2.4773 -1.0523  0.2452
#>   0.9210 -1.2193  0.6214 -0.7716 -0.4043
#> 
#> (3,.,.) = 
#>   1.1330  0.1184 -1.2779  1.2477  1.3106
#>   0.2376  0.5517 -0.8407  1.1091  1.1702
#>  -1.0238 -1.6408  0.6979 -0.6028  0.0034
#> 
#> (4,.,.) = 
#>  -0.4956  1.7282  1.0025 -0.5227  0.3847
#>   0.4199  4.1740 -2.8427 -2.4136  0.4493
#>   0.9146 -2.0954  0.6234  1.4543 -1.8229
#> 
#> (5,.,.) = 
#>  -1.7857  0.1301  0.2353 -1.2195 -1.5819
#>  -2.1918  0.5546 -1.1934 -0.4523 -2.9275
#>   5.7949  1.6360 -5.6127  3.0694  4.3485
#> 
#> (6,.,.) = 
#>  -0.9521 -0.2892 -0.9584 -0.6583  0.2249
#>   4.8173  1.2223 -0.5714  5.6072 -1.1272
#>  -2.5567 -1.4235 -3.1505 -2.9103  0.5152
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]