Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.8395  0.6548 -3.1898 -0.8541  0.8061
#>  -1.6630  0.3591 -1.9247 -0.4442  0.3496
#>  -3.8866  1.1354 -3.7220 -2.0572 -2.0071
#> 
#> (2,.,.) = 
#>  -1.2374  0.0556  3.9538  4.5286  0.6594
#>   1.6513 -1.7249 -1.5692 -3.0905 -0.5269
#>   1.9210 -1.7827  0.4263 -0.2179 -0.1896
#> 
#> (3,.,.) = 
#>  -2.2509  1.7289 -0.8659  1.3615 -1.9034
#>   0.5570 -2.1288  0.3435 -5.4526 -1.5581
#>   0.2928  0.7789 -0.6583  1.7149  0.3021
#> 
#> (4,.,.) = 
#>   0.7964  1.5510 -1.8961 -0.4981  2.7901
#>  -1.1201 -1.4044  1.5185  0.5697 -1.9918
#>   2.7276  1.0756  0.1661  0.3419 -1.8206
#> 
#> (5,.,.) = 
#>  -0.8309 -0.0314  0.3407 -0.4262  0.6787
#>  -1.1601 -1.3739  1.7689  0.7956 -0.7334
#>   0.6799  1.7215 -0.2932 -1.6619 -1.6643
#> 
#> (6,.,.) = 
#>   0.3990 -0.3641 -0.9008  0.9094  1.0345
#>  -0.3381  0.3203 -0.2776 -0.2806  0.2005
#>   0.4854 -1.9989 -1.2633 -1.1591 -4.2017
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]