Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   1.1198 -0.8426 -2.7041  3.0640 -7.4867
#>   0.7764  1.0872 -1.1128 -0.1979 -0.6333
#>   0.4748 -3.1463 -1.2201 -1.0460 -2.2943
#> 
#> (2,.,.) = 
#>  -2.4257 -1.7318 -0.7675  0.0688  4.0999
#>   0.9964  2.2610  0.2690  0.0332 -1.6553
#>  -1.5498  1.2618 -0.7911 -0.7210 -0.0398
#> 
#> (3,.,.) = 
#>   0.2591 -0.5382  0.5062 -0.4297  0.2034
#>   1.3329  1.5723 -1.2321 -2.2231 -0.4926
#>   1.4927  1.9828  1.1133 -0.3019 -3.1039
#> 
#> (4,.,.) = 
#>   1.4229  0.1720  4.5083 -2.1911 -0.1704
#>  -2.0360  3.4486  0.2418  1.5800  0.6538
#>  -3.5240  0.0432  2.9879 -2.3285 -0.4023
#> 
#> (5,.,.) = 
#>  -1.5634  0.5047  0.6353  0.4098 -1.2184
#>  -0.7397  0.0288  1.4815  1.6063 -1.1775
#>  -0.2238 -0.3595  0.7893  0.3122 -0.6645
#> 
#> (6,.,.) = 
#>  -0.0891 -0.5239 -1.5097  0.4909 -0.1035
#>   0.1928 -0.3296  1.7291  1.9860 -0.0233
#>   1.5897  0.9203 -0.3295  1.0713  0.5669
#> 
#> (7,.,.) = 
#>   2.1096 -1.2987  0.8066  4.3063  2.6556
#>  -1.5708  2.6923 -0.2546 -1.3083 -0.9567
#>  -0.5981 -0.8744 -0.5043 -0.3600  1.0038
#> 
#> (8,.,.) = 
#>  -4.2770  1.5581 -1.2074  1.4418 -1.0335
#>   0.7274 -1.3362  0.7641  1.1194 -0.5482
#>   1.4040  1.2002 -1.3740 -1.7677 -1.4025
#> 
#> (9,.,.) = 
#>   0.7124 -0.5661 -1.6922  2.3203 -1.9373
#>  -1.3343 -1.1475  1.8946 -1.0692  3.1012
#>   0.7237  1.5791 -1.8007  0.0056 -2.0635
#> 
#> (10,.,.) = 
#>   2.5923  1.8049  0.9713 -0.8139  0.8445
#>  -5.6130  3.5439 -1.5127  1.5044  2.5344
#>   3.0378 -3.4945  6.2923  0.9006 -2.9564
#> [ CPUFloatType{10,3,5} ]