Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   0.0660 -0.1954  2.1124 -1.2080  1.1224
#>  -0.1897 -0.7522  0.0844 -0.7424  0.5702
#>   1.1638 -0.4303  0.0711 -0.4668  1.2379
#> 
#> (2,.,.) = 
#>  -0.4015 -0.3909 -0.6132 -0.5402 -1.7972
#>  -0.5917 -0.4539 -0.4613 -1.0180 -2.7934
#>  -0.5479  1.0278 -1.2994  0.0315 -0.3472
#> 
#> (3,.,.) = 
#>   1.1895  0.8930 -2.0019 -2.4518  3.0547
#>  -4.7869  3.0762 -0.4592  2.3440 -0.9095
#>  -4.4427  1.7928  0.1188  2.3876 -1.4518
#> 
#> (4,.,.) = 
#>   1.1390  7.9713 -4.3375  0.2528 -0.5951
#>   0.5961  2.9722 -2.5726  1.1647 -1.2111
#>  -2.5959  3.6598 -3.8002 -1.6282  2.4066
#> 
#> (5,.,.) = 
#>  -1.1526  2.7543  6.9721 -1.3597 -1.9611
#>   0.8414  2.6504 -0.8246  4.3008 -1.6977
#>   1.3018 -4.4361 -7.3296  0.4284  2.0089
#> 
#> (6,.,.) = 
#>   0.8388 -0.3617 -2.4322  7.5410 -1.8262
#>   0.9390 -1.1162 -1.8366 -1.2175  0.2905
#>   2.1770 -0.3751  1.1407  4.3036 -0.0044
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]