Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.6334  0.5639  0.8522 -0.3269  0.4519
#>   2.3227  0.7907  1.8081 -0.6827  2.3247
#>  -1.9887 -2.1092 -1.8243  0.9310 -2.7618
#> 
#> (2,.,.) = 
#>  -4.2586  1.7049  1.2347  0.4508  2.1493
#>   1.4271  2.1126 -4.3277 -1.8381 -7.2499
#>  -1.4551  1.1198 -1.6926 -0.4475 -3.0904
#> 
#> (3,.,.) = 
#>  -1.2234 -2.2359  1.9594  1.7947  0.6398
#>  -0.1205 -1.6869  0.0955  1.7883 -0.3419
#>   0.5780  0.6031 -0.6584 -0.4574 -0.4253
#> 
#> (4,.,.) = 
#>  -0.8226  0.4590  1.7143  0.9004  1.0757
#>  -0.5532  0.8365  1.1837  1.1249  0.0032
#>  -1.4326  0.0607  0.1917  1.6200 -0.0292
#> 
#> (5,.,.) = 
#>  -5.0915  1.5256 -3.5425  2.0963  0.9736
#>   0.1807  0.1679 -1.9563 -0.6433 -1.4241
#>  -0.6161 -3.8406  1.2591  0.4446  0.7101
#> 
#> (6,.,.) = 
#>   0.2764 -0.9004  0.2992  4.2003 -2.6710
#>   1.0951 -0.8287 -2.6795 -0.7071 -0.2255
#>   2.0024  1.1142  0.1536  2.2307  1.6929
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]