Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -5.0635  1.9392  3.9461  0.6367 -0.8225
#>   0.1180  2.4890  2.8089 -0.9155  0.8499
#>  -0.8159  2.3431  1.6395 -0.7424  1.1169
#> 
#> (2,.,.) = 
#>   0.3425  0.1086 -0.2058  0.6819 -0.4364
#>   0.0346  1.9280 -0.9041  0.2396  1.9449
#>  -0.3511 -0.6824 -0.0181 -0.4356  0.5061
#> 
#> (3,.,.) = 
#>  -0.1755 -0.5339 -5.2626 -0.3231 -3.0918
#>   1.4014 -0.9242  3.5397  0.9513  1.1646
#>   2.2320  1.6403  0.4558 -1.5050 -2.0610
#> 
#> (4,.,.) = 
#>   0.4526  0.4739 -2.2753  2.3272 -0.9386
#>   0.7649 -0.9365 -0.0743 -1.0446  0.8995
#>   1.1937  1.0277 -2.5615  2.4137  0.6937
#> 
#> (5,.,.) = 
#>  -3.0257  1.3485  2.5725  6.7797  1.1665
#>   5.1540 -1.5547 -1.3166  5.4847 -0.1915
#>  -0.1901  0.8920 -1.1290  0.2358 -0.6780
#> 
#> (6,.,.) = 
#>  -2.7709 -2.2483 -4.6865  2.2100  2.3471
#>   0.5858 -0.5140  0.3905  0.2371 -0.7983
#>   0.6269  1.5785  1.9933 -0.0135  0.7261
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]