Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.1285 -0.6315  0.2992  1.7540 -1.4087
#>   0.5725  0.6278 -0.7170  0.8631  0.3787
#>   1.4624  1.1031 -0.9418  2.2704 -0.8357
#> 
#> (2,.,.) = 
#>   0.1871 -0.5152 -1.5862  0.7154  0.6934
#>   0.4718  3.2718  4.0258 -1.9492 -1.0483
#>   0.5195 -0.8540 -4.4962  2.1324  1.3302
#> 
#> (3,.,.) = 
#>   1.2696  0.5261  0.5893 -1.0633  3.4743
#>  -0.5246  1.1113  3.0405  1.1833  4.9946
#>  -1.5152  2.7321  2.2741  2.6284 -0.3865
#> 
#> (4,.,.) = 
#>  -1.5319  0.5099  0.3240 -0.4765  0.0950
#>  -0.7761 -1.1860  1.6262 -1.1357  1.7381
#>   2.1939 -0.5003 -2.3219  1.2959 -2.9804
#> 
#> (5,.,.) = 
#>  -1.0128 -1.1415 -0.3832  1.6283 -2.2393
#>  -1.4975  0.3393  0.0341  0.0717 -0.8423
#>  -0.5619  1.3286 -2.5042 -2.3193  0.5128
#> 
#> (6,.,.) = 
#>  -1.0412  0.3834  1.6908  0.8837 -1.0569
#>   0.9128 -1.4494  0.4319  0.0944 -0.4874
#>   0.8687  0.1751  0.1917 -0.5162  1.1689
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]