Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   1.2073  0.4710  2.3071 -0.2159  1.5429
#>  -0.0827  1.3981  0.9944 -2.1732 -0.9045
#>  -0.3879  1.7342  1.0035 -1.3264 -1.9876
#> 
#> (2,.,.) = 
#>  -0.5097  3.1217 -0.2428 -1.7233 -1.5944
#>   0.3960  1.3717 -0.3402 -0.4010 -0.5335
#>  -0.1741 -3.1865 -0.2380 -0.3291  0.9962
#> 
#> (3,.,.) = 
#>  -3.2884  2.8286  1.9227  5.9582  0.7336
#>   0.0107  2.9952  3.5590  4.2408 -2.6941
#>   1.0636 -1.6141  0.6819 -4.3874 -1.4394
#> 
#> (4,.,.) = 
#>   0.2724  0.4036 -1.0200  1.5231 -0.1341
#>  -0.7892  0.4070  0.1569  2.7432 -0.6627
#>  -0.5307  0.3420  0.3067  2.0958  0.0527
#> 
#> (5,.,.) = 
#>   0.6130  0.4330  0.9657  0.1611 -0.3409
#>   1.2149 -2.1799  2.2003  1.3170 -0.2345
#>  -0.8365 -0.1169 -1.3945  1.0876 -0.3946
#> 
#> (6,.,.) = 
#>   1.9442 -1.6866  2.5728 -0.3739  0.4020
#>  -4.1875 -0.6698 -0.3523  1.4369  0.2336
#>  -0.0377  0.0486  0.1859  0.3410  0.8216
#> 
#> (7,.,.) = 
#>   0.6071  0.9916  2.2655  0.5970  0.3372
#>  -2.5206  0.0829  1.0446  0.0509 -0.2445
#>   1.4686 -0.0879 -0.9168 -4.1588 -0.4562
#> 
#> (8,.,.) = 
#>   2.0457 -0.0096 -3.9321  4.0868 -0.1263
#>   2.0941 -1.3904 -1.3416  4.2638 -1.3134
#>   2.2427 -0.7598 -2.4538  3.9653 -0.2193
#> 
#> (9,.,.) = 
#>   0.4030 -0.8426 -1.0297 -1.3059 -0.1121
#>   0.5411  0.3338  0.3922  1.3703 -0.0814
#>   1.2788 -0.8245 -2.0546 -0.7158  1.5336
#> 
#> (10,.,.) = 
#>  -2.7292 -2.4858 -0.6313 -1.0950  0.3862
#>   0.3818  1.2517 -1.6538  1.4522 -1.2155
#>  -0.2296 -0.6718 -0.3928  0.0545 -2.5712
#> [ CPUFloatType{10,3,5} ]