Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   1.0939 -0.7311 -0.0731  0.5278  0.0286
#>   1.0431  0.5636 -0.8905 -0.2208 -0.0767
#>  -0.4048 -4.0079  4.5800  1.5405  0.8139
#> 
#> (2,.,.) = 
#>   1.7872  0.9021 -0.8171  0.2503 -1.3543
#>  -1.8266  0.7323 -0.9535  2.1583 -0.3096
#>  -0.8082 -0.4202  0.0784  0.2791  0.1574
#> 
#> (3,.,.) = 
#>  -3.9134 -0.6465 -1.7334 -1.1108 -0.0514
#>  -0.2742 -1.3032  1.1194  1.2378  1.7762
#>  -2.1122  0.7500 -0.8803  0.3229  0.2809
#> 
#> (4,.,.) = 
#>  -3.5439 -3.7472  1.5700 -2.8792 -0.0361
#>   0.5181 -1.8100  1.0373 -1.5125  0.4696
#>   1.2850  2.4925 -2.9605 -2.2392  0.9122
#> 
#> (5,.,.) = 
#>  -0.8317  0.3632 -2.4859 -0.2724  1.1668
#>   1.1181 -1.5565  1.7028  0.7390 -0.4336
#>  -0.2592 -0.8861 -1.7184 -0.4880  0.7530
#> 
#> (6,.,.) = 
#>  -0.0865  0.5331 -4.1490 -1.7660  2.2339
#>  -1.1060 -0.7065 -4.3225 -2.8366  2.0782
#>  -1.7176  0.0949  0.1018 -0.5917  0.6514
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]