Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   0.1778 -0.6783  0.4014 -0.1425  1.5257
#>  -0.8578  0.7537 -0.0670 -0.1042 -1.5066
#>   2.2421 -1.2869 -1.0429  0.8074  0.7369
#> 
#> (2,.,.) = 
#>  -2.8605  1.2006  2.1226 -2.1945  0.0656
#>  -3.8395 -0.3904  1.8906 -0.7796  0.3894
#>  -3.0807 -1.2156  0.3562  1.1904 -0.2674
#> 
#> (3,.,.) = 
#>  -0.4065  3.7476  2.1430 -0.7780 -2.5267
#>   0.2350  1.0248 -0.4875 -2.8008 -1.6540
#>   2.0081 -0.6027  4.1736 -2.0865  1.5173
#> 
#> (4,.,.) = 
#>  -1.1484 -1.3360  0.7140 -0.6685 -1.4459
#>  -2.1234 -1.2098  2.2302 -1.2912 -0.8460
#>   0.6522 -0.9168 -1.5834 -0.1336 -3.9270
#> 
#> (5,.,.) = 
#>   0.4978 -1.7823 -0.8774  0.6251 -0.3141
#>   0.6733 -2.5101  0.4055  6.2004  2.0490
#>   0.4295  2.1334  3.1090  0.5767  0.9026
#> 
#> (6,.,.) = 
#>   0.7283 -0.1061 -4.1386  0.1443 -0.5784
#>  -1.9299 -1.3717  5.1432  4.5905  0.7182
#>   2.4363  2.5116 -3.7656 -2.4887  1.3780
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]