Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.6089  0.7583  0.0221  0.1114 -2.4457
#>  -0.0955 -2.8358 -0.3426  0.3381  1.9085
#>  -1.8720  3.7818 -0.1815 -2.2802  0.7733
#> 
#> (2,.,.) = 
#>   3.6451 -1.7869 -1.7027 -3.4108  0.2783
#>  -0.4414  0.8224 -1.2181 -1.0183 -1.5944
#>   2.0036  4.4918 -0.2467 -2.2535 -2.6574
#> 
#> (3,.,.) = 
#>  -3.1203 -1.0905  2.9438  2.4864  2.2981
#>  -5.7933 -3.9976 -0.5665  4.0996  3.0325
#>  -1.9311 -0.1222  3.1074  1.7614  2.6500
#> 
#> (4,.,.) = 
#>  -2.7511  6.3000  6.1487  1.8039 -1.6682
#>  -3.1825  0.8671  2.4185 -1.2110  0.1296
#>  -3.7383  2.3279  2.1158 -1.4233 -3.4724
#> 
#> (5,.,.) = 
#>  -1.6059  1.2675 -0.7710 -0.1506 -2.2079
#>   3.6206 -1.5299 -1.6236  1.5885  2.2352
#>   4.9128  0.2166 -1.6538  2.1428 -0.8684
#> 
#> (6,.,.) = 
#>   1.6043 -0.8843 -1.2804 -0.7652 -2.2668
#>   1.3870 -0.2086 -0.7732 -0.4910  0.0594
#>  -4.0801  2.0416  2.8121  1.1169  1.6387
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]