Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   2.4251 -5.5041  0.4991  1.1405  0.4017
#>  -0.6899  0.5897 -0.1874  1.6158  0.9308
#>   0.2193  3.2226  0.3820  0.1504  0.8884
#> 
#> (2,.,.) = 
#>  -4.0667  1.4788 -0.9049 -3.2656  2.9442
#>   5.6466 -2.3931  1.7541  3.0594 -4.0756
#>   1.8122  2.1531 -0.0194  0.2037 -1.5452
#> 
#> (3,.,.) = 
#>  -1.3192  1.1300  1.4752  1.4848  0.5819
#>   0.0549  0.3239  1.6421  1.5671 -1.6459
#>   0.4043 -0.8158  2.1304 -0.5556 -0.7915
#> 
#> (4,.,.) = 
#>  -0.2427 -0.4034 -0.4335 -0.1840  0.0725
#>  -3.6666  1.2592  1.1780 -1.3917 -1.3460
#>  -1.1347  0.4137  0.1278  0.0919  0.5855
#> 
#> (5,.,.) = 
#>  -1.1396  1.0359 -2.4591  1.8142 -1.1836
#>  -1.2689 -0.2376 -2.3407  2.0830 -0.2805
#>  -1.7899  2.3071  1.3721  1.0859 -0.6364
#> 
#> (6,.,.) = 
#>  -1.6512  0.9108  0.4630 -0.6530  0.1111
#>  -0.1347 -1.0060  0.2168  1.0445 -0.9127
#>   0.2359  0.7632  0.2209  0.8836 -1.0241
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]