Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.4889 -2.0104 -1.1046 -1.3955 -2.4930
#>  -1.7219 -0.6197 -1.1800  2.4371  1.5690
#>   0.0976  1.3237 -0.0877  3.8749  3.6165
#> 
#> (2,.,.) = 
#>   3.9657 -2.4170  4.3128 -2.4996 -0.7504
#>   1.4614 -0.7155 -0.0160 -0.3413 -1.9054
#>   1.3607 -0.0243 -4.1370  3.8807 -0.2193
#> 
#> (3,.,.) = 
#>  -1.7205  2.1428  2.2024  2.0447 -2.1403
#>   0.1456  1.1348  0.0447  1.8662  0.2117
#>   2.7881  0.4264  0.2275 -3.3427 -1.6395
#> 
#> (4,.,.) = 
#>  -0.8459  2.3916 -0.6650 -0.2867 -3.7055
#>  -0.3100 -0.2873  1.0112  1.5415  3.7825
#>  -0.0348  2.0677  1.1561  0.0687  0.3390
#> 
#> (5,.,.) = 
#>   1.3961 -0.0417 -4.4403 -2.1520 -1.1897
#>  -2.7663  6.1144  4.8615 -3.0061 -0.8445
#>   1.8515 -1.4026 -0.0719  1.4831  2.0266
#> 
#> (6,.,.) = 
#>  -1.3151  1.3057 -0.2119  0.1864  0.2287
#>   0.5269 -2.9284  1.9270 -0.7315  0.8400
#>  -1.4972  0.0632  1.4383 -0.0125  0.9570
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]