Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.3214 -1.6777  1.2081  1.1399  0.9637
#>   4.5108  4.5281 -4.8498 -3.0039 -0.9605
#>  -1.3295 -2.2433  0.8005  1.4100  1.3069
#> 
#> (2,.,.) = 
#>  -1.1518 -0.0588 -0.8485 -0.1759  0.6231
#>   0.2310 -2.6407  2.1988  0.5730 -0.6402
#>   0.6209  1.2139  0.2995  0.0163  0.2548
#> 
#> (3,.,.) = 
#>  -0.8811 -1.2228  2.0733  0.4210 -1.9389
#>   0.6008  2.3299 -5.2377 -6.5269  0.0126
#>  -0.9746  0.6698 -2.0701  0.0319 -0.6551
#> 
#> (4,.,.) = 
#>   0.7236  1.3351  0.0261 -0.9584 -1.5442
#>   3.0993 -1.2652 -1.1133 -1.2430 -0.5982
#>   2.0621 -0.0178 -1.2716  0.7677 -0.9081
#> 
#> (5,.,.) = 
#>   2.9225 -2.0675 -1.6620  1.8267 -1.7529
#>  -1.2001  0.2853  1.3209 -0.1075  1.3479
#>   0.7580 -0.6245  1.4765 -3.8496 -1.7173
#> 
#> (6,.,.) = 
#>  -2.5386  0.5903  0.7015  1.2761 -0.2109
#>  -0.9819 -2.2500  0.4763 -1.3482 -2.8448
#>  -2.1463  0.1970  1.3118  0.1138 -0.9769
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]