Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor 
#> (1,.,.) = 
#>   3.2041  2.8227 -1.8704 -2.0817 -0.3630
#>   1.6176  3.7683 -0.6770  1.6203  2.2567
#>   0.2043  4.4495  0.5935 -1.1354  1.3342
#> 
#> (2,.,.) = 
#>   0.1009  0.6787 -0.4399  0.1653  0.2866
#>  -3.9584  4.2686 -1.2125 -0.3897 -0.9442
#>   1.0453  3.6925  0.4842  1.5266  0.7501
#> 
#> (3,.,.) = 
#>   4.0569  2.3061  2.1652 -3.9326 -2.7924
#>   1.5600 -1.7567  0.0872  1.1937  2.0629
#>   0.0528  0.2691  2.0843  1.5672  0.1863
#> 
#> (4,.,.) = 
#>   3.9869  1.6668  1.2100  0.7417 -2.6102
#>  -1.1215 -0.1443 -0.1286 -0.0321  0.1667
#>  -0.2521  2.1209 -1.1180  0.3102 -1.4409
#> 
#> (5,.,.) = 
#>  -0.2530  1.6106  0.8054 -0.3099 -2.7916
#>   0.2185 -0.7362  0.1725 -0.5402  1.5874
#>   4.2868 -0.8454  3.5697  2.0019  4.8906
#> 
#> (6,.,.) = 
#>  -1.5340  0.5558  1.8805 -2.4926 -2.7061
#>  -0.5991  0.4464  0.6274 -0.3571  0.5673
#>   0.1491  0.1695  0.8655  0.2694  1.0192
#> 
#> (7,.,.) = 
#>  -0.3517 -0.2261 -1.7849  0.2723 -0.7295
#>   2.3300  2.4134 -0.0452 -0.2731  1.1269
#>  -0.2934  0.9943 -0.2145 -1.1562  1.4052
#> 
#> (8,.,.) = 
#>   0.2641 -0.8123  0.3338  2.7016 -2.6568
#>   3.7162  0.8339 -1.0552 -0.4369 -1.3411
#>   1.2215  1.6794 -1.2193 -1.8392 -2.8012
#> 
#> (9,.,.) = 
#>   0.4360 -1.7670 -1.7761 -0.9085  0.4833
#>   1.4944  2.7179 -1.4438 -0.0510  0.6708
#>  -0.2843 -0.0329  0.6219 -0.0111  0.8767
#> 
#> (10,.,.) = 
#>  -1.7485  0.9969 -2.1181 -1.0870  0.1361
#>  -4.9081 -1.8396  3.8518  1.4138 -3.1362
#>  -0.2592 -2.0966  2.1822 -1.6956 -2.2890
#> [ CPUFloatType{10,3,5} ]