Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   3.6742 -2.3355 -0.0977 -2.5926 -1.6718
#>   2.7647  0.5696 -0.1814 -0.3948 -1.4333
#>  -0.6821 -0.3532  1.1814  0.6419  0.2100
#> 
#> (2,.,.) = 
#>  -3.8499 -3.8044 -0.9711 -0.9240 -1.3479
#>   6.4764  2.8914  0.7360  3.4391  1.7101
#>  -1.4817 -0.0502  0.8253  0.0135  1.1383
#> 
#> (3,.,.) = 
#>  -0.3886 -0.0753  0.8334  0.3110 -1.4781
#>   0.5472  0.3911 -1.0614 -1.6150 -0.2432
#>  -0.0484  0.6808 -1.3165 -1.1292 -0.5953
#> 
#> (4,.,.) = 
#>   0.9008 -0.0340 -0.1857  0.1573 -0.5953
#>   3.4784  1.3015  2.5000  0.1417  1.8770
#>  -1.4684  0.3927  0.7829  0.4265  0.4525
#> 
#> (5,.,.) = 
#>   2.8372  0.2024 -0.1940 -0.7536  0.9819
#>   0.4763 -0.5945  0.4089  1.4280  3.0084
#>  -2.5628 -0.1057  0.0711  0.2358 -1.9828
#> 
#> (6,.,.) = 
#>   0.3852  0.7714 -0.6180 -0.7516  4.3105
#>   0.8500  2.4670  1.4616 -2.3089  0.2261
#>  -1.6871 -4.8703  0.9582  3.9142 -2.4020
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]