Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>   2.0800 -0.6436  0.9451 -0.4325  0.6095
#>   2.4588 -1.7842  2.2168  0.2463  0.1785
#>   0.1918 -1.4434  4.6763  5.5388 -1.0127
#> 
#> (2,.,.) = 
#>   2.2045  2.0192  0.8324 -1.8229 -0.6264
#>  -1.5069 -1.7047 -2.3760  2.4360  0.2382
#>  -0.5601 -1.8909 -0.3444  3.3301  0.6616
#> 
#> (3,.,.) = 
#>  -2.8704  4.6951 -0.3059  0.0398 -0.1316
#>   1.3066 -1.6354 -0.4059 -0.2260  0.4911
#>  -0.4115 -3.3861  0.6507  3.9147 -1.5694
#> 
#> (4,.,.) = 
#>   0.2171 -0.6638  1.0188 -1.0694  1.9055
#>   3.1253 -0.6315  5.1626 -2.8841 -1.2703
#>  -0.0773  1.1587  1.1679  0.0971 -1.7772
#> 
#> (5,.,.) = 
#>  -0.3172  1.0833  0.4888 -0.4663 -3.6673
#>   0.5417 -0.0013  1.9440  1.4921  2.6561
#>  -0.8837 -0.1154  1.6663  1.5848  4.8048
#> 
#> (6,.,.) = 
#>   0.9930  0.4361  3.6160 -2.2722  1.0712
#>  -1.6186 -1.5184 -1.8254  0.3612 -0.1022
#>   4.3592 -0.0639  1.9566  1.2625  2.4508
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]