Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>   1.0210  2.7049 -0.9664  0.7957  0.1293
#>  -0.4933 -1.8630  2.3455  0.6151  3.8926
#>  -1.0849 -3.1045  0.7438  0.8001 -0.1579
#> 
#> (2,.,.) = 
#>   2.3964 -2.7891  0.8204 -1.1800  0.9056
#>   2.7371 -2.9209  2.3177 -1.0667  2.4906
#>   1.8928 -2.3640  1.7216 -0.9403  1.3233
#> 
#> (3,.,.) = 
#>   0.3448  1.3283 -0.8018 -0.9369  0.3992
#>  -1.7148  2.3491 -1.5569 -0.7127 -1.3244
#>   1.9539 -3.4013  1.9035  1.8585  1.7428
#> 
#> (4,.,.) = 
#>  -0.0764  1.8403 -2.7692  2.5194 -2.0861
#>   1.2301 -4.5982  0.8069 -3.3820  1.3046
#>   0.0897  0.7369 -0.9891  1.6830 -1.4267
#> 
#> (5,.,.) = 
#>  -0.9270  1.2921 -2.2950  1.2501  2.5073
#>  -1.2230  1.5349 -1.2191  1.2667  0.4208
#>   0.0373 -1.1896 -0.2869  0.2923  0.2681
#> 
#> (6,.,.) = 
#>   1.4030 -1.8163 -0.1938  1.7823  4.4009
#>   0.6856 -0.0745 -0.2161  1.6600 -0.2662
#>  -0.9233  0.1278  0.1477 -1.9701 -0.1352
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]