Bmm

Bmm

Usage

torch_bmm(self, mat2)

Arguments

self

(Tensor) the first batch of matrices to be multiplied

mat2

(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}
#> torch_tensor
#> (1,.,.) = 
#>  -7.3269 -0.9055 -0.1712 -2.8253  0.0374
#>  -2.8233  2.1665  0.2947  1.7298  0.8067
#>  -0.1271 -0.8098 -0.2658 -0.4607 -0.2093
#> 
#> (2,.,.) = 
#>   0.8927  2.0244  6.9167 -0.6882  6.0454
#>   1.2272  2.2596  2.1708 -0.0544  1.9610
#>  -1.5006  0.0242 -0.4426  0.3590 -3.0851
#> 
#> (3,.,.) = 
#>  -2.6211 -3.5706 -0.7782 -1.2627  0.8767
#>  -0.7425 -0.6695 -0.2951 -2.0124  0.4376
#>   0.3705 -1.2623 -0.6407 -0.8524 -0.9748
#> 
#> (4,.,.) = 
#>  -0.6912 -0.1504 -0.4806 -2.6621  0.3076
#>   1.1913  0.9973  0.1199 -0.5646  1.3443
#>   1.0452  1.3320 -0.0468 -2.4069  2.0044
#> 
#> (5,.,.) = 
#>   2.0561  0.3868 -0.8075  0.6019 -0.1519
#>   0.2067  0.0873 -0.1638 -2.7550 -2.1336
#>  -0.4323  0.1922  1.3842 -0.5422 -1.2405
#> 
#> (6,.,.) = 
#>   0.3536  0.6129 -0.1428 -0.4540 -2.2400
#>   0.5213 -2.0869  4.1341  3.7226  3.6221
#>   0.4532  0.1380 -1.9463 -2.4928 -2.1449
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]