Bmm

Bmm

torch_bmm(self, mat2)

Arguments

self
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=NULL) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \mbox{out}_i = \mbox{input}_i \mathbin{@} \mbox{mat2}_i $$

Examples

if (torch_is_installed()) {

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.6211 -0.8530  1.1944  0.8516  1.1184
#>  -2.2018 -4.1880  4.5416 -1.6691 -1.1600
#>   3.1887  4.1504 -0.4820 -2.2705 -0.3774
#> 
#> (2,.,.) = 
#>   0.2455 -0.4173  0.8786 -0.2117  0.6592
#>   1.7411 -1.9232  0.1743  0.5288  2.2229
#>   0.7046  0.9903 -0.9157  1.6101 -0.5121
#> 
#> (3,.,.) = 
#>   1.2033  1.1039 -0.3670 -0.5569  0.4407
#>   0.7555  1.3261  0.6735 -0.9865  0.3012
#>   0.5181  0.3776  0.7521 -1.3174  1.8878
#> 
#> (4,.,.) = 
#>   5.7031  2.9813 -2.1782 -3.1892  6.3813
#>  -0.7057  1.0582  2.0760 -1.5039 -1.2297
#>  -1.9532  6.8996 -0.9820  1.3495 -0.8939
#> 
#> (5,.,.) = 
#>   7.0527  2.3998  0.0215 -5.7693  0.2135
#>   2.1896  0.2817  0.0562  1.0070  0.4186
#>  -0.8100 -0.4415  1.7517  1.3142 -1.3024
#> 
#> (6,.,.) = 
#>  -0.2663 -1.7206 -0.2007 -0.7503 -0.2975
#>  -1.7147 -0.8785 -0.4750 -1.6726 -2.2941
#>   0.0973  0.5447  0.3498 -0.6119 -0.3221
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]