Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   5.6892 -1.5348 -1.5379  2.8267 -3.6770
#>   2.2482 -4.6566  4.8787 -5.4262 -2.1089
#>   1.8893  0.3297  9.9385  1.2329 -2.6758
#>  -2.3877 -5.0071  5.7513 -4.4861 -2.6696
#>  -2.0648 -3.6349  2.8652  0.1436 -1.0787
#> 
#> (1,2,.,.) = 
#>   0.6132 -8.4510  0.6498 -0.3583 -4.3749
#>   2.1860  7.6954 -1.8916  3.1816 -8.6062
#>   2.4533 -8.9973  7.6363  3.7075  6.0385
#>  -2.2297  4.1373 -1.8167 -4.5809  5.4271
#>  -4.4248  4.4120  1.1890 -0.8920 -5.3349
#> 
#> (1,3,.,.) = 
#>  -10.4060   9.9321   6.1824  -0.3022   2.3050
#>    0.2609   5.1091   1.3600  10.9270   4.1668
#>   -5.3851  -8.3338  -7.5453 -12.6673 -14.3939
#>   -7.4250   5.5745  -4.7382   1.7632   1.3804
#>    9.3654  -2.5703  -1.4898   9.1332   6.5417
#> 
#> (1,4,.,.) = 
#>   5.1823  0.6931 -1.1300 -0.9603 -2.0170
#>   4.4448 -2.7152 -0.1939 -1.6305  1.6599
#>  -0.0493  2.2116  4.4826 -4.1084  3.9941
#>  -1.2820 -3.4386  9.0678 -1.8736  0.1246
#>   0.3805 -5.7181  3.8209 -6.2933 -2.7474
#> 
#> (1,5,.,.) = 
#>    3.4750   0.3996  -1.2262   0.6942  -3.1152
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]