Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -0.4397 -9.0573  4.3113 -5.8397  0.5880
#>   9.6840 -7.1660  9.2082  6.6264 -0.1278
#>  -1.3842  6.2694  0.3305 -0.1288  0.2399
#>   6.0055 -4.7041 -5.9832 -0.6751  0.8072
#>   3.5275  7.0111  0.0420 -2.6260  1.2327
#> 
#> (1,2,.,.) = 
#>  -5.4465 -4.2311  1.2463  0.7084  2.8307
#>  -2.2898 -8.3693 -5.9230  9.5325  1.5926
#>   1.4395  3.8530  4.3828 -0.8894  1.4216
#>  -1.1063 -3.2820 -8.1310 -7.4024 -2.2240
#>  -3.7557 -7.3674 -1.6312  0.2443 -1.6224
#> 
#> (1,3,.,.) = 
#>   -0.2266  -2.8177   2.2373  -0.4564  -0.6191
#>   -1.3993   1.8321  10.7725  -8.6401   1.6026
#>    7.7804  -0.7695   4.0465   1.4844   1.3363
#>   -4.4599   0.2824  -5.0820   0.4548   6.0957
#>    6.5398   1.5100  -7.2163  -0.0963   3.4941
#> 
#> (1,4,.,.) = 
#>  -3.6622 -3.8021 -6.8499  1.8513  2.1711
#>   1.9116 -3.8014  3.0835  3.1660  3.2589
#>   7.9916  5.5507  0.9761 -0.9976 -3.3420
#>  -1.3105 -7.4696 -0.0484 -1.2300 -3.3902
#>   4.9958  4.6108 -0.9643 -3.7406 -0.3229
#> 
#> (1,5,.,.) = 
#>  -0.0366 -7.0048 -0.1786  4.3493 -4.0699
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]