Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -3.7618 -2.8736 -9.2148 -6.6090 -3.5067
#>  -5.6329 -4.6232 -0.3692  6.5187  5.4131
#>  -8.3279  0.8426 -5.1779 -1.4464  4.5519
#>  -4.9392  6.2546  1.7455 -5.7514  4.4651
#>  -1.2054  5.6463 -0.0543  4.6659 -0.1963
#> 
#> (1,2,.,.) = 
#>   1.2311  8.3756 -2.0206 -5.7291 -3.4018
#>  -3.7318  4.5338 -2.4982  3.4560 -3.0050
#>  -2.9482  5.3006  2.1578  8.8709  4.5118
#>   0.3317 -2.9929 -0.0748 -6.8214  5.3966
#>   2.0014 -3.4242  0.2209 -6.0943 -0.4674
#> 
#> (1,3,.,.) = 
#>   -1.3244   3.5514 -11.5351   6.6500   0.4750
#>   -5.5835   3.7662  -4.0095  -0.2583  -2.3032
#>   -2.4630  -3.3032   2.8783   3.6557   0.0843
#>    0.3948  -0.5085 -10.4507  10.9461  -5.9370
#>    4.0357   6.6734  -1.3054   3.7732  -2.0024
#> 
#> (1,4,.,.) = 
#>  -3.0311  7.2341 -12.3554  2.1120 -4.4841
#>   4.5259 -2.0022  0.3612 -1.3015  8.4026
#>  -7.5662  7.2091 -9.5903  3.0573  4.5585
#>  -5.3043 -4.4705  1.7020 -7.7206 -5.6085
#>  -4.8963  0.0035 -9.3369  2.1505 -3.9532
#> 
#> (1,5,.,.) = 
#>   -3.0815  -0.0790   1.8822  -2.3418   7.4299
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]