Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    0.4879  -2.1485  -6.8267  -6.1572  -1.1605
#>    1.0563  -0.5533  -0.6431  -3.0841  -5.9350
#>    2.0543   9.3461 -11.4231  -6.0030   1.6456
#>   -6.0278  -9.4142  11.1499  -5.3660   4.4964
#>   -2.7011  -5.2003   1.2260  -1.0008  -1.0961
#> 
#> (1,2,.,.) = 
#>   1.7491 -3.8301 -1.1652  1.9346  4.2625
#>   1.2420  7.5213  2.7211 -7.4333 -3.9275
#>   0.6463  4.5628 -11.1726 -0.5006 -1.2437
#>   4.9345 -4.7966  9.0361  5.2418  5.7537
#>   4.3222 -5.0561 -7.3920  6.2413 -8.4864
#> 
#> (1,3,.,.) = 
#>   -2.2364  10.7371   5.4683   1.3107   6.3084
#>    5.8999  -2.9050   0.8024   0.6744   1.3088
#>    1.4385   0.5883  -4.6247   9.4727  -3.3075
#>   -3.4503  11.7512  -3.2358  -3.3017   3.3358
#>    1.1321  -4.4055  -7.4665   3.3149  -2.6719
#> 
#> (1,4,.,.) = 
#>    7.3633   4.3069   5.6274   4.7738  -1.0048
#>   -3.3816  -5.0371   6.6014   1.2482   1.4431
#>    4.9248   9.9627  13.2288  -7.4177   5.6305
#>   14.4306   2.1907  -2.3225   0.8835  -5.2659
#>   -0.0414   8.0166  -4.1801  10.0646   1.5638
#> 
#> (1,5,.,.) = 
#>   1.9647 -9.1581  8.2419 -2.1599  1.6115
#>  -1.8827 -2.7748 -2.2860 -2.7845 -7.9056
#>   7.8042  4.3473 -1.8834  0.1686  1.0674
#>   1.2445 -7.7169  5.3805 -5.0261 -0.0844
#>  -0.4985  2.8594 -2.5773 -3.8701 -2.4184
#> 
#> (1,6,.,.) = 
#>    4.8412  -2.3151  -1.6490 -12.4626   3.8354
#>    2.4242  -0.4316  14.6355  -5.3826   0.4273
#>   -3.6817  11.3297  -7.1475   8.0963   5.0572
#>    1.9084   0.0226  -6.9541  13.4873   0.4756
#>   -4.0841  -0.4066   5.4161  -2.1339   3.4137
#> 
#> (1,7,.,.) = 
#>    3.1099   1.1146  -2.5114   8.3860   2.1656
#>   -8.0951  10.1284   0.4130   1.3060   2.1638
#>    8.2317   0.0773   7.5524  -8.4309  -3.8893
#>    3.7730   5.1695   3.7212   2.4577   4.6923
#>   -2.2331   6.5874   2.1668  -2.6344  -1.3919
#> 
#> (1,8,.,.) = 
#>   -1.5626  -4.9009  -5.9740   3.6430  -5.2546
#>    4.4833  11.8366   1.1611   9.1714   4.9276
#>   -3.3200   0.4647   2.1372  -7.1005   5.9299
#>    9.0098 -13.6726   0.3690  -6.8959   1.2197
#>    0.6564  -1.9087  -4.5195  13.3130  -2.6067
#> [ CPUFloatType{1,8,5,5} ]