Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   3.8147  7.2519  4.2807 -0.0165  6.2830
#>   0.6280 -4.5107 -7.9426 -1.0316  0.2559
#>   0.8196  2.5432  7.0180 -7.1401 -6.4127
#>   8.0116  3.8627  2.3061 -4.7825 -1.8484
#>  -3.5593 -8.5827  4.9659 -3.6820  3.2831
#> 
#> (1,2,.,.) = 
#>  -0.7630 -3.4677  0.2892 -2.5910 -2.8903
#>  -0.0313 -0.1987  6.0915 -5.2171 -3.5472
#>  -0.1217 -1.2482 -1.6079  0.2196 -2.2115
#>  -5.3390  1.2412  0.8940 -2.2555 -4.3209
#>   2.7169  0.8984 -5.8739 -1.5867 -4.4739
#> 
#> (1,3,.,.) = 
#>    2.6839  10.7330 -16.9346  -2.7955   2.8746
#>   -9.1723  -9.9001  -4.8586   2.3773  -8.1902
#>    0.0713  11.8308  -4.6873  -1.7034   1.9965
#>  -16.2549  -9.9807   3.7185  -7.8499   1.7166
#>    3.4866   5.8017 -15.9992   1.4134   4.9402
#> 
#> (1,4,.,.) = 
#>   0.3974 -4.6059 -1.3491 -2.9137 -11.0089
#>   1.4604  3.3745  9.9717  9.4335  0.4737
#>  -6.8909 -6.8539  7.3830 -4.4516  6.1151
#>   7.9377 -0.4806 -16.2468  2.7294 -0.5453
#>   6.0768  4.2061  4.6463  2.6617  1.1456
#> 
#> (1,5,.,.) = 
#>    0.0580  -5.0374  -2.4547  -4.3945  -0.2598
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]