Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -2.7228  2.3589  2.3867 -4.6176 -1.4823
#>   0.0693 -0.0213 -13.9902  2.9183  2.5367
#>  -0.6884 -3.1010 -5.5232 -0.1166 -6.5811
#>   2.8819  7.6899  4.0807  1.7607  0.6784
#>   5.4739  0.7398 -4.9664  5.6071  0.6354
#> 
#> (1,2,.,.) = 
#>   3.4738 -6.8159 -0.8677  1.6705 -6.4856
#>  -0.4549  1.8630 -3.8548 -8.7778 -6.3493
#>   3.8670 -1.6193  6.3032  4.2974  2.4551
#>   3.3778 -7.1657 -1.3640  1.6781  0.3385
#>   3.1118  3.4975  8.7177 -3.8029 -2.6049
#> 
#> (1,3,.,.) = 
#>  -0.4877 -1.2097 -3.7460  5.5587  2.3634
#>  -2.2123  6.1670 -8.8966 -6.9087  5.5580
#>  -2.6519  5.4993 -7.9944  4.3739  3.9776
#>   2.5871  2.4238 -6.4521  7.0301 -7.6353
#>  -0.4826  6.3034 -3.3411 -2.8633 -0.2149
#> 
#> (1,4,.,.) = 
#>  -5.2780 -4.0221  3.2990 -1.7458 -2.9960
#>   2.6470 -4.0485 -9.4247  4.8141 -0.1523
#>   6.5061 -1.5030  4.9370  1.6339 -1.5311
#>   6.0780 -2.5913  4.3047 -5.1592  0.9584
#>   1.7433  4.3082 -4.5114 -0.0629  3.7653
#> 
#> (1,5,.,.) = 
#>   2.8825  2.4330 -2.2081  2.2530 -0.0212
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]