Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    7.6214  -7.6481  -4.3437   7.2091  -8.0786
#>   -2.8115  11.6526  -5.8083   0.4237  -6.1810
#>    6.0592 -11.5629   3.4804   1.6731  -1.0441
#>   -6.1169   7.9233   7.1900  -6.0277  -2.5971
#>  -10.6264   5.6973  -3.8280  -2.9920   1.9355
#> 
#> (1,2,.,.) = 
#>    4.1957  -8.8336   5.1948  -0.5569  -4.9197
#>   -4.8187  -0.2698  -5.3156  10.5117   4.4808
#>    1.8768   2.6863 -10.8682  -1.1260  -2.7745
#>    0.5117  -3.1635  -6.5087  -4.5724  -8.5313
#>    2.0538   0.4118   1.6625   1.5000  -1.6333
#> 
#> (1,3,.,.) = 
#>  -3.5512 -5.6644 -0.8496  0.9435  2.6590
#>   1.5676  1.6154  2.9518 -11.2213 -3.6539
#>  -4.4003  8.0123  1.2513 -4.2955  0.4824
#>  -3.3075 -7.2776  0.0470 -0.0490  2.6572
#>   0.2839 -6.7541 -8.6655  1.1291 -2.2550
#> 
#> (1,4,.,.) = 
#>   -5.3165 -10.7524   1.0283  -0.0846   6.1547
#>   -2.1207  -0.5489  -3.6413   8.0220   6.6669
#>   -1.9319   4.1201   5.4321 -22.3296  -4.2119
#>    3.6146  -0.5872  -2.8993   0.6250  -1.2636
#>   -2.7925  -3.1517  -1.0734  -1.6149  11.9258
#> 
#> (1,5,.,.) = 
#>  -2.3458  9.1166 -6.8078 -0.3170  0.6975
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]