Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -5.1331  4.7797 -4.0680 -4.2020 -6.8022
#>  -2.9640 -4.0775  7.1061 -9.2752  0.6932
#>  -8.6455  1.6912  6.7678 -0.1760  2.5245
#>  -5.5380 -4.1889  5.2903 -10.6547  4.1971
#>  -0.0813 -2.6960 -2.2956 -2.8546 -2.1600
#> 
#> (1,2,.,.) = 
#>    7.0649  -4.0309  -8.4625  -9.3043   1.6112
#>   -0.9433   0.5856  -2.7866  -3.5224  -8.9487
#>   11.5914  -3.9858   0.7878   1.4717   2.7960
#>   -4.3855  12.6908   3.7747  11.7825   2.4629
#>   -0.7472  -2.0165  -6.9971  -3.6127   1.8080
#> 
#> (1,3,.,.) = 
#>   0.0563  4.0559 -0.4620  3.2196  0.1130
#>  -4.2301  1.2358 -0.3342  4.3785 -2.4997
#>  -5.3483 -1.6519  1.9072  4.2035  3.6878
#>  -1.3687 -2.0283 -5.2053  4.1307  5.8216
#>  -0.6652 -2.6487 -5.0690 -4.0412  7.8879
#> 
#> (1,4,.,.) = 
#>    4.7381   5.1686  -8.6121  -2.2475   3.9694
#>    3.0547   3.3629   1.4733  -2.7324  11.9589
#>    4.4838   6.4125  -7.3643  -1.6398  -0.5613
#>    6.4113  -5.3147  -2.2068  -3.1736  -0.0860
#>   -1.1479   2.6599   5.8599   0.0348  -2.8166
#> 
#> (1,5,.,.) = 
#>  -1.1393  6.3376 -5.2927  3.0278 -0.0053
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]