Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   0.6210 -4.7270 -2.9902 -6.0124 -5.5233
#>   5.2061 -7.8686  6.2608  1.6086 -0.5556
#>   4.4541  1.2610  1.7899  3.9458 -1.6240
#>   1.9666  0.1961 -4.9692  5.7203  6.0919
#>   0.7481  5.6244 -6.7180  1.8431 -0.7305
#> 
#> (1,2,.,.) = 
#>   2.0055 -5.5151  4.3711 -0.4255 -0.6842
#>  -0.9935 -0.2619  4.7177 -7.1754 -0.6981
#>   3.2067  8.5048  2.4475  0.5775  0.2361
#>  -4.2766  3.8686 -2.9924  0.2539  1.0668
#>  -7.1152  4.9333 -3.2763  2.1677 -4.3526
#> 
#> (1,3,.,.) = 
#>    3.8535  -6.5680   5.1813   0.2828  -1.0325
#>   -8.6711  -9.1708  -5.0071  -8.7308   0.9417
#>   -0.0678   0.1218  -3.2315   1.0576   3.8740
#>    2.8940  13.0429  -0.0477   3.5292  -4.8019
#>   -8.4222   5.7520  -2.5985  12.5487  -0.5936
#> 
#> (1,4,.,.) = 
#>    4.7443   8.1788   2.2830   5.9759   2.2516
#>   -5.7397   3.1215 -12.9859   0.4823  -1.2528
#>    3.5337   5.0199   4.0074  10.8468  -0.0362
#>   -2.1184   2.9107   1.3186  -6.4830  -0.3331
#>    4.6260  -3.4369   6.4570  -3.2865   3.2675
#> 
#> (1,5,.,.) = 
#>   -4.6355   0.5333   1.7989   0.0236  -0.8543
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]