Conv2d

Conv2d

Arguments

input
NA input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
NA filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
NA optional bias tensor of shape \((\mbox{out\_channels})\). Default: None

stride
NA the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
NA implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
NA the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
NA split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=None, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See ~torch.nn.Conv2d for details and output shape.

.. include:: cudnn_deterministic.rst

Examples

# \dontrun{

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)

#> torch_tensor 
#> (1,1,.,.) = 
#>  -1.2050  3.5601 -7.8325 -0.6184 -5.9300
#>   6.9992  0.3900  1.3901 -2.5574  6.5135
#>  -2.3490  3.9786  4.1985 -9.6240  8.3693
#>   0.1293  0.5186 -2.7322  9.4996 -1.1448
#>  -5.2164 -12.5999  8.9623 -8.9888  7.3018
#> 
#> (1,2,.,.) = 
#>  -0.8450 -8.2603 -2.9019 -5.5598 -4.9509
#>  -5.6410  1.2666 -5.8790  2.1525 -1.3928
#>  -0.2024  2.1010  6.3255 -0.5656 -2.5797
#>  -0.3285  6.7826  5.8021 -1.7078  7.7191
#>  -0.0922 -8.5321 -5.4983  0.7221 -3.5291
#> 
#> (1,3,.,.) = 
#>   2.5094  9.2626  5.0731 -3.6366  3.3318
#>  -0.2142  2.5216 -0.6932  8.3285 -0.4724
#>  -2.7548 -2.3855  1.6016 -9.5280 -0.9662
#>   1.6404 -7.5112  0.5470 -3.7868  0.9206
#>   0.0776 -1.8698  3.2148 -2.8353 -1.9503
#> 
#> (1,4,.,.) = 
#>   -1.0442  10.8465   5.7719   2.8028   0.3426
#>   -1.9633   3.7087  -3.1929  -2.0919  -8.2913
#>   -0.9261  -3.8192   9.3410  -9.4944   8.6766
#>   -5.9383  -2.2824  -4.0878 -10.5887  -2.2831
#>    6.7283   2.9823   0.6066   0.9479  -1.2099
#> 
#> (1,5,.,.) = 
#>  -3.4877 -1.5824 -6.4475 -1.1290 -2.3183
#>  -2.2229 -5.1762  9.2085 -3.5268 -3.8315
#>  -6.3696  2.3858 -12.1739 -2.6510 -3.0961
#>   4.4813  3.8243 -0.4043  2.4245 -0.3913
#>   0.6648 -2.4358  0.4671 -3.7840 -3.5538
#> 
#> (1,6,.,.) = 
#>   0.8825  0.3176 -6.6245 -6.0051  2.4430
#>  -2.0265 -0.9768  6.7438 -7.4899  2.0777
#>  -6.8935 -2.2333 -3.3011  2.0151  2.5786
#>   0.2469 -4.8563  2.3537  0.5968  3.6600
#>  -1.4781  0.2248 -3.1730 -6.7766 -3.0347
#> 
#> (1,7,.,.) = 
#>   1.5889 -4.5167  0.0906  7.6198  1.8365
#>   5.0933 -2.3551 -2.6166 -7.5808 -2.0514
#>   4.0381  6.9844 -2.3409  2.5503  6.2012
#>   6.1860  8.5588  0.8248  1.6048 -4.7339
#>  -2.7566  3.9998  8.0454 -1.6114 -0.9366
#> 
#> (1,8,.,.) = 
#>   4.3153 -3.2445  0.4090  1.5033  6.4462
#>   6.4586  2.8058 -3.3607  5.9048 -2.4407
#>  -3.9344  4.6081  5.3624 -0.2189 -0.3718
#>  -4.3796  5.2653 -0.4628  8.5329 -6.1232
#>   0.7405  5.8740  5.0339 -0.6710  0.8794
#> [ CPUFloatType{1,8,5,5} ]

# }