Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   -3.5486  -7.1217  -3.2540  -5.9799   1.1629
#>   -4.3378   2.2990 -16.4566   9.0228  -5.5777
#>   -2.6722  -9.2867   4.0511  -4.4372  10.4134
#>   -4.7778  -0.7326  -1.6019  12.0401  -3.1893
#>    0.2123  -0.5914   1.1211   3.0260   6.8859
#> 
#> (1,2,.,.) = 
#>   -3.5645  -0.6611   1.0267   0.0105  -2.2313
#>   -4.8099  10.2621   0.1669   0.9807  -2.2823
#>    0.7857   0.4305   0.4022   7.5398   0.2195
#>   -1.5292   7.5032  10.2237  -4.5421  -3.0606
#>    4.4675   1.8375   9.2117  -1.7632  -6.9053
#> 
#> (1,3,.,.) = 
#>   -4.0560  -3.3721  -3.9417   1.2809   1.6676
#>    3.7261   1.6556  -7.7465  -0.6448   3.6757
#>   -1.8130  -5.0859  -3.2322   4.3222   5.7618
#>   -2.3173  -7.4043   2.5139  15.0216   0.5374
#>   -1.4209  -1.0366   0.8718   4.0616   3.1472
#> 
#> (1,4,.,.) = 
#>  -9.5668  3.2701  6.5622  9.6684  2.5335
#>  -4.4349  3.4779 -1.2737  2.9751  1.7985
#>   2.6595  7.2363  4.4882  0.3747  1.3252
#>   5.3966 -4.0692 -2.2576  9.7051 -0.5021
#>  -0.2109  4.6618  0.3269 -6.7229 -2.3171
#> 
#> (1,5,.,.) = 
#>  -6.8665 -0.7471  3.0223  5.9422  0.0773
#>   0.1951  6.6746  4.5183  0.2309 -3.8129
#>   4.9832  2.5449 -5.4966  3.8797  3.9093
#>  -1.2040 -1.1130  7.1904  3.6236  3.7172
#>  -1.3379 -5.7016  6.7872  7.4608  4.0352
#> 
#> (1,6,.,.) = 
#>   -3.9338   0.3756  -3.0094  -0.7198   3.5040
#>    0.3583  -1.8715  -4.4996  11.1578  11.3811
#>   -7.6832  -2.1231  13.6712   7.9567   4.1675
#>    7.7207   2.5612   9.3426   5.4903  -9.0039
#>   -1.2806   1.9558   6.0637  -3.7703   2.7139
#> 
#> (1,7,.,.) = 
#>   -3.5015  -2.8558  -3.0872   2.0101   8.1843
#>   -2.7114  -4.0357   0.2954   1.3045   2.9680
#>   -0.0320   0.3198  -2.8986  -0.1796  -4.9480
#>    6.8524  -9.7840  11.4722   0.6094  -8.8633
#>   -2.3049  -2.9029  -4.5625  -5.2812  -1.8420
#> 
#> (1,8,.,.) = 
#>    4.1203  -6.8209  -7.9792   0.5355  -6.1896
#>    8.1814 -14.2806   2.7375 -18.4326  10.9079
#>   -0.8891  -8.0756   0.7869  10.5435   0.2842
#>  -11.1754   5.4648   9.2422   8.0793  -3.8445
#>    3.1076   1.0749  -4.5827   4.5922  -1.6976
#> [ CPUFloatType{1,8,5,5} ]