Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -1.2037  2.8533  3.0966  2.7602  7.2992
#>   6.6099  0.8040  2.1820 -2.2556  1.7042
#>   8.9937 -2.8627  7.9067  2.2703 -0.2219
#>  -1.0960 -2.0544  9.6476 -4.1084 -1.7429
#>   2.2151 -0.6322 -10.7968 -2.1303  0.9502
#> 
#> (1,2,.,.) = 
#>   -0.5524 -10.7146  11.9610  -1.8760   2.8073
#>   -1.4550   5.9693   1.4488   0.5483   0.5193
#>    4.7319  14.7168  -1.9500   3.4680  -3.1870
#>    2.3077  -8.0817   5.5639   2.1393  -4.4990
#>   -0.0674  -0.8898  -4.0815  -8.3260   0.0084
#> 
#> (1,3,.,.) = 
#>  -6.0152 -1.1616  6.4675 -10.9283 -1.7122
#>   2.0931 -3.9504 -5.1826 -5.1925 -2.1788
#>  -1.9279 -5.6841  3.6513 -4.5984  1.9415
#>  -2.6984  2.3140  5.6889  3.2366 -1.6959
#>  -1.2875 -5.5225  2.4667  1.2794  0.6366
#> 
#> (1,4,.,.) = 
#>   -1.3725 -11.9432   6.6814  -2.0626 -12.8413
#>    2.7790  10.4186  16.5919   3.1043   1.6991
#>   -2.1525  -0.6591  -7.1645  -2.4902  -2.3716
#>    4.5753  -0.9017  -5.4827   3.2854   1.9315
#>    0.5009   4.8452   3.5895  -1.3763  -5.0205
#> 
#> (1,5,.,.) = 
#>   -2.4641  -4.1384   0.0252   5.7390  -3.1903
#>  -11.9943  -5.2914   1.6707  -5.6204  -5.7324
#>   11.1295   6.3085  -4.1165   0.6665   6.0104
#>   -2.1664  -4.9256  -2.1528   6.1870   8.8261
#>   -2.8591  -4.4202   0.0518   1.5891   0.1274
#> 
#> (1,6,.,.) = 
#>    7.7084   7.0540  -6.6758   6.2570   1.1702
#>    6.1343  11.9692  -0.8624   0.9134   7.8712
#>   -4.3698 -10.2884  -2.7717  -5.6516   0.0050
#>    4.3435   2.9247   0.0498 -12.5641  -1.2759
#>    3.1506   7.0418  11.2416   3.0961   1.4600
#> 
#> (1,7,.,.) = 
#>    8.8728 -10.0021  -1.0044   7.5667   2.3943
#>   -5.8614   9.1914 -11.8036   1.8316   7.5042
#>   -6.9303  11.0166  -4.5717   3.7854   4.8000
#>    5.1606  -8.6169   1.7790  10.5232  -1.9409
#>   -0.5102  -2.8458   3.5851  -7.8170  -1.8686
#> 
#> (1,8,.,.) = 
#>  -1.2384  1.3963 -4.1241  7.3818 -3.9786
#>  -1.8505  1.5575 -2.4711  1.1911 -7.8127
#>  -4.5371  2.7213 -7.5125 -5.1210 -6.9012
#>   3.0100 -9.5409 -0.6372 -9.0458  0.6989
#>  -9.1543  4.1365 -7.5186  2.5672  7.7889
#> [ CPUFloatType{1,8,5,5} ]