Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -1.5553  -2.3483   7.6456   6.5007  -0.3473
#>   -8.6220   3.4330   3.9425  -6.9423   0.9180
#>    6.3151  -6.1122   3.6371  -7.2665  10.2832
#>   -1.6446   1.3575  -4.6583   0.7745  -0.8268
#>    2.4858   0.0763   2.3942   0.5526  -4.0451
#> 
#> (1,2,.,.) = 
#>  -0.7907 -5.4746  3.1452 -4.6013 -0.3526
#>  -0.0790 -0.1334  5.8237 -0.8588  2.3477
#>  -4.5662  2.5850  0.3865 -11.0322  2.2622
#>   9.6151  7.7557  3.1338 -3.9470 -1.0708
#>  -6.5531 -1.5014 -3.1337  5.4262 -0.0273
#> 
#> (1,3,.,.) = 
#>    0.9873   0.2783  -1.6370  -6.2729   3.1244
#>   -5.6118  -0.3992  10.7749  -2.0045   8.0761
#>   -0.6057   4.7651  -4.3037  -7.6184   0.0957
#>   -1.9538   1.9499 -14.8640  -0.9127   3.7142
#>    5.9992   5.0179  -0.2612  -4.5648   1.2170
#> 
#> (1,4,.,.) = 
#>   1.5804 -2.2261  3.1328  2.1154 -2.9721
#>  -0.3420 -8.7395 -2.5239 -8.6153  3.6052
#>  -1.0489 -0.7044  5.5746  3.9232  0.2373
#>  -3.2464  5.2771 -8.5470  1.6731 -6.9801
#>   1.2986 -1.7310  9.3044  0.2758  7.0516
#> 
#> (1,5,.,.) = 
#>   5.0170  5.7143 -7.3290 -0.1282  1.3319
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]