Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    3.8569   0.2329  -2.4751  -2.4023  -9.0091
#>  -10.5521  -3.4191   1.0158   3.6790 -13.6014
#>    2.1764  -4.2928   8.0815  -7.7557  14.6866
#>    7.6970  14.6685  -4.0728  -1.5431 -11.8510
#>    0.4240  -6.3825   4.1172  -0.6481   2.1638
#> 
#> (1,2,.,.) = 
#>  -7.1201  1.7817  6.1472  4.6471  1.3535
#>   6.7903  1.8407  5.8802 -1.9514  3.7789
#>   2.5248 -3.7805 -0.2298 -0.2656 -0.9014
#>  -1.0995  3.4218  4.5275  2.6735 -7.4065
#>  -2.6218  4.5126  4.4163  0.7770  0.5402
#> 
#> (1,3,.,.) = 
#>  -0.9075 -1.3425  3.2155 -3.7205  3.1044
#>   6.2013  2.6372  1.3184 -12.5262  8.4944
#>   0.0260 -2.8125 -1.7484  3.1217  3.1035
#>   7.3364 -0.8105  1.9614  1.5870  0.1104
#>   1.6171  3.4342  1.5984 -3.9596  4.0404
#> 
#> (1,4,.,.) = 
#>   -2.3104   2.3278   4.4693  10.9861   1.2399
#>   -1.7717   5.0541   0.8358  -3.7936  -7.0112
#>   -3.0243   2.6510   1.0832   2.6181   1.9369
#>    0.9276   3.4044  -3.7765  -8.9957  -2.8890
#>   -0.8776   5.0554   1.4160   1.2946  -0.0465
#> 
#> (1,5,.,.) = 
#>   1.5639 -1.9991  3.2091  0.3871 -0.7073
#>   4.1515 -4.6491  3.3199 -3.1584  0.9138
#>   4.6945  2.8303  6.7107 -0.6192 -0.4188
#>  -0.3976 -3.7313 -3.6691 -2.9883 -2.5855
#>   1.6160 -7.4333  4.1836 -4.3755 -2.5735
#> 
#> (1,6,.,.) = 
#>   3.5371  4.9186  6.1973  4.0796  5.1086
#>  -8.2944 -4.2939 -14.3125  0.1191 -10.0509
#>  -4.3509  1.2977  2.8589  0.9587  6.2226
#>  -4.0990 -5.7003  8.0716 -2.6998  6.5902
#>   2.8760  4.6510 -5.4740  5.8864 -1.5556
#> 
#> (1,7,.,.) = 
#>   -7.2738  -3.3895   1.1254   1.3818  -7.3193
#>    3.5279  -7.9514  -5.3414  -6.3420   3.1169
#>   -2.8624  -5.5002   6.0441  -5.8398   4.4587
#>    0.0431   4.8956  -4.9912   0.1093  -4.7037
#>    9.0929   3.4590  10.1236  -4.6726  -0.2556
#> 
#> (1,8,.,.) = 
#>   -3.2478  -3.8540  -1.0620   4.2485   3.1979
#>   -0.9923   8.1070   3.4919  -0.0049  -2.7287
#>   -5.0899  -4.1093  -1.5783  -6.2119   1.8222
#>    3.7956   3.9447  -0.6606  -6.1794   5.8117
#>    2.8509  12.9711   6.2676  -3.2129  -4.6998
#> [ CPUFloatType{1,8,5,5} ]