Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -6.3262  -3.1745   6.8439  -0.2147  -0.7471
#>   -1.1311   8.5408   6.4542  -4.7346   1.9577
#>    2.1446  17.7405   4.6555   0.2107   3.9928
#>    2.3831  -8.2887  -7.1573  -8.8190   2.6731
#>   -8.5725 -11.3091  -8.6246   2.0091   2.9034
#> 
#> (1,2,.,.) = 
#>   -1.7503  -7.9661  -9.2161   1.5285   1.0738
#>   -6.6069   0.7948 -10.1998  -1.0614  -7.1272
#>   -3.4033   4.7531   5.3908  14.5483   8.1443
#>    9.3285  -0.7142   5.0458   0.4961  -6.6820
#>   -7.9288  -5.3265  -2.5698  -3.8538   1.1888
#> 
#> (1,3,.,.) = 
#>   -0.0699  -5.0774   9.6797  -7.2533  -2.0257
#>   -4.1320   1.6490  14.4432 -11.3671  -0.3361
#>    1.2511  -0.1385   1.3726  -8.6758   7.7033
#>    0.3243   3.0028   1.6191  -4.6294   2.4779
#>    1.3193  -5.0803  -2.0904   3.3670   1.1712
#> 
#> (1,4,.,.) = 
#>   -3.8773   6.4254   3.3204   5.8551   4.0348
#>    3.5473  10.2587 -11.0564  -0.4985  -2.0603
#>  -10.6730  -5.6151  -7.0121  -2.7681   1.0559
#>   -6.9054   7.0037 -10.1233   0.4418   2.7211
#>    0.8239  -6.6099   5.7825  -1.7503  -0.0935
#> 
#> (1,5,.,.) = 
#>    4.9686  -2.1027  12.8052  -2.0316   8.2605
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]