Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -2.9705  0.6919 -3.8448 -0.3075  2.8986
#>  -1.2170  3.6623 -2.7800  3.5377  0.5290
#>  -1.6757 -7.0444  0.4438 -1.5843  8.0627
#>   1.0366 -10.4036  5.6271  2.8769 -0.9213
#>  -4.3760 -3.5317  2.8569  3.4470  5.7636
#> 
#> (1,2,.,.) = 
#>   0.2274  3.0856  6.4797 -5.2177  1.1317
#>   6.0625  1.1112 -6.9719 -0.7155  1.7873
#>  -5.4734 -1.4407 -5.5922  6.5090  2.1762
#>   0.0693 -2.4782 -5.2749  1.5426  3.3766
#>  -2.8258  1.4044 -4.5733  4.5010 -0.5219
#> 
#> (1,3,.,.) = 
#>   1.4686  3.0412  7.0236  4.5024  0.3074
#>  -1.3171  5.0028  6.3376 -0.6539 -4.4066
#>   1.3381  1.8425  7.1578 -4.2934 -0.5961
#>   6.8297  1.9485  6.2502 -6.6090 -1.2034
#>   1.8723  3.0213  2.6075 -3.4112 -2.0823
#> 
#> (1,4,.,.) = 
#>   2.5501 -4.8259 -5.7225  3.1932  1.7182
#>   0.1031  4.9959 -2.1084 -7.1863 -5.2037
#>   3.6435  0.2082 -0.7920 -2.0804 -3.5658
#>   5.1210 -2.4327 -4.2196  0.0292  0.1519
#>  -4.4608  0.7043 -12.3139  4.4498 -0.3211
#> 
#> (1,5,.,.) = 
#>   -2.2489 -14.5275  -4.8270  -4.4639   3.0554
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]