Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    6.1912   3.1512  -3.9837  -6.0357  -3.2990
#>   -1.4917   0.1511   1.9908   6.5183   1.3646
#>   -6.6152  -1.8614   2.7919   1.1652  -0.1021
#>    6.6967   1.0255  -1.3240  15.1735  -8.9362
#>    4.0116  -1.6908  -4.1517  -4.3438   7.4023
#> 
#> (1,2,.,.) = 
#>    2.4800  -4.7913  -1.6520   3.9213  -0.2119
#>   13.7255 -10.6737  -8.8504   2.0383  -4.5083
#>   -2.4294   2.7447  -2.5893  -6.3582   8.8948
#>   -3.4666   0.2759  -1.0724   2.3053  -5.1092
#>    5.3472   9.8207   3.3989  -0.4521  -1.7494
#> 
#> (1,3,.,.) = 
#>    1.3002   4.1023   0.1749  12.2213  -1.9836
#>   -1.5754  -4.0261   1.5128   6.1798  -1.5561
#>    6.2795  -5.7921   0.4592   0.7556  -1.0426
#>   -7.8286  -0.9130   0.8890  -6.2132   5.0615
#>    0.9437  -6.1572   5.8409   0.6821  -4.6155
#> 
#> (1,4,.,.) = 
#>   -1.9780   9.1973   9.3558  -9.2889   2.0103
#>    2.2864   3.5985   3.8848  -1.2723  13.8569
#>   -0.0816   1.7572   5.4910   8.3117  -2.6265
#>    2.5719  -0.0734  -3.3429   7.8817  -2.3823
#>   -0.3837  -0.5298   0.0701  -2.4968  -5.4668
#> 
#> (1,5,.,.) = 
#>    1.6813   8.1098  -5.8593  -7.8758   2.6861
#>    1.3376  -4.2494   8.4932   1.7570  -2.7101
#>   11.2011   2.4364   8.0352  -1.4855  -1.5757
#>    3.7617   9.1308  -5.1691   8.4753  -2.2933
#>   -1.8262  -6.6776   0.0706   0.4150   3.4934
#> 
#> (1,6,.,.) = 
#>   -5.9285   6.3511  -6.0420  -3.6994  -0.8923
#>    1.4884   6.5877 -10.4279  -1.6800  -3.6066
#>   -3.2911  12.8681  -1.2398   4.9871   7.7141
#>   -3.9247   9.3752  -1.8777   0.6430  -1.5830
#>    0.2597  -1.8668   4.2603   3.4733   2.4945
#> 
#> (1,7,.,.) = 
#>    4.3621   4.0451  -8.0065   5.4915   3.1457
#>   -3.1675   4.0963   2.7828  -4.9746  -3.2921
#>   11.7568  -8.2425  -2.6722  -3.4414  -0.7506
#>   -1.5973   0.1580   1.2271  -2.1056   2.7798
#>   -5.5817   2.6926   5.0657  -0.2080  -2.2711
#> 
#> (1,8,.,.) = 
#>   1.7182  4.1062  0.6233  2.4717  5.0801
#>   6.5988  4.2647 -0.4290 -1.7904 -3.3128
#>  -1.3016  1.6924  5.1291 -0.9635  1.2795
#>   9.9448 -2.8539 -3.9430 -5.4761  0.8236
#>   5.8087  9.5697  3.1718 -4.3638  2.5913
#> [ CPUFloatType{1,8,5,5} ]