Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -2.2272   0.1579  -1.8472  11.5603   2.3288
#>   -3.2061  -4.5486  -2.3270 -13.8779  -7.7861
#>    4.7108  -0.2098  -5.1171  -0.7684   1.0906
#>   -4.3943 -13.1048  -6.3079  -3.5968   2.4073
#>    1.5180   7.2845   0.6028  -5.3092   0.0380
#> 
#> (1,2,.,.) = 
#>   -0.7560  -0.9425   6.4240  -3.2309   5.3288
#>    0.0442   5.7597  -1.6277   2.5243  15.9118
#>    5.5002  -6.8342  18.6938  -4.5236  -4.0378
#>   -0.2076  14.4670  -3.2933  -2.3203  -0.9623
#>   -4.8024  -4.6648  -5.7470   1.5515  -1.4522
#> 
#> (1,3,.,.) = 
#>    1.6583   5.4161  -4.0544  -0.3446   6.4042
#>    4.8568   0.9479  -7.3299   5.7656  10.9523
#>    6.5624   0.3732   6.8542  -6.3341   1.4088
#>    5.0483  -0.7374  -0.1272   0.0549  -3.9619
#>   -5.4080   2.2435   3.1119  -5.2096  -5.5669
#> 
#> (1,4,.,.) = 
#>  -1.5112  4.4659  0.0878 -1.2094  0.4470
#>  -2.7772 -10.2252  4.4123  0.9056  0.8880
#>   2.3495 -1.3826  9.9115 -1.0979 -3.8633
#>  -5.0889 -5.1910 -5.8273  4.0040  7.8209
#>  -5.6176 -3.5455 -7.9369 -3.7642  2.5973
#> 
#> (1,5,.,.) = 
#>   -4.6552   4.4757 -14.3778   2.5082   2.7293
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]