Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -8.5353 -15.1513  -6.8071  -7.1383   0.6963
#>    5.9454  -2.7262  -2.4483  -2.8073  -7.0181
#>    2.6521   5.5140   3.1206   0.8119  -5.8976
#>    6.8712   1.1140  -5.9485  -2.9822   4.5346
#>    2.9707  11.3264  -4.1129  -6.1829  -3.0461
#> 
#> (1,2,.,.) = 
#>    7.0739  -1.4328   5.0690  -6.6876   0.6078
#>   -4.6167  -1.6313  -5.3573  10.4210  -6.9401
#>    1.9730  -2.8903   8.2092  -0.9084  -0.2206
#>   -3.7881   0.7487  10.8822   0.6511  -6.8203
#>    7.1310   2.3071   0.5469   1.8876  -0.9895
#> 
#> (1,3,.,.) = 
#>   1.7629  0.9127  0.6901 -2.6033 -3.6840
#>   5.3270  5.5561 -2.9616  6.9330 -0.4614
#>   4.1066 -0.6250  5.4168 -0.4083  3.4633
#>   9.0672 -1.4065 -1.9251  3.1501 -2.2638
#>   0.9213  3.0609  2.8656  1.8844  0.2026
#> 
#> (1,4,.,.) = 
#>   1.7771 -0.4597  2.5922  2.2850 -0.0619
#>   2.0869 -3.6327 -2.4941 -1.4589  2.2180
#>  -2.6531 -0.0516 -5.2757  0.6926  6.3825
#>   0.4204 -1.3913  9.4159 -1.8461  2.3028
#>  -1.2944 -2.5482 -1.0503  2.8823  0.0373
#> 
#> (1,5,.,.) = 
#>   1.1051  0.2849  3.0912 -1.3357  1.3236
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]