Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    0.9330   5.9163  -3.7955  -6.7919  -1.8877
#>    1.1751  -1.2969  -6.4483   4.1286   3.9723
#>    5.2779   6.0618  -2.0892  -1.1818   1.9467
#>    1.1437   0.1813  -7.8647  -5.7195  -4.5994
#>    4.6020  11.2504   7.6711  -3.3272   2.6438
#> 
#> (1,2,.,.) = 
#>   -1.2248   4.2196   7.5841   2.8989  -0.4751
#>    5.5587   3.7955   4.0788  -8.1019   0.3894
#>    2.5300   5.9224   1.8242  -1.7954  -0.5497
#>    3.3506  11.9531  -0.2167  14.4222   0.3198
#>   -4.6914  -0.7929  -2.7473  -0.0024   3.3916
#> 
#> (1,3,.,.) = 
#>    3.4866   3.2288  -5.6834   1.9738  -4.7371
#>    3.3467   4.8392  -0.1283   0.4005  -0.7439
#>    0.7762  -1.2675 -11.4759  10.0584   9.9336
#>    8.3771 -11.6595  -4.0543  -7.5377  -4.0770
#>   -3.8217  -8.6571   8.1904  -2.2053   4.4371
#> 
#> (1,4,.,.) = 
#>    1.2007  10.1325   4.5137   2.7677  -4.2161
#>    0.6021   1.6251  -4.3985  -0.0645   2.6852
#>    3.5178  -5.1669 -10.8532  11.1787   1.0899
#>    9.9297  -2.1577   4.9247  -7.1082   4.6139
#>   -3.3512  -5.4665  -5.7081   6.4850   0.1357
#> 
#> (1,5,.,.) = 
#>  -2.5979  0.2886 -5.4850 -0.3849 -0.4538
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]