Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    6.0273   5.8360   4.2001  -2.9789   4.6401
#>    0.7793  -5.8630  -8.2795  -5.6532   5.7866
#>   -5.7940  -0.6959  -2.1899  -4.9562  -3.6739
#>   -4.1689  -2.1111   2.1359  -6.5465  10.4177
#>    1.3321  -2.9974   9.9332  -0.1753   1.3983
#> 
#> (1,2,.,.) = 
#>    4.4744   9.7590  -2.6591   3.3579   5.0566
#>    1.7507  -4.9441  -2.0675   4.2160  -4.8944
#>   -5.2388   1.1833   4.3375  -0.5160   2.2510
#>    3.8457  11.0895   2.2529   4.6892   6.9889
#>    2.0916   3.5938   2.6947   0.8206   1.6253
#> 
#> (1,3,.,.) = 
#>  -0.9300 -5.1464 -5.6456 -3.1695  2.1679
#>  -6.1456 -3.0466 -12.1117  6.9943  5.4393
#>  -14.9906 -5.4630 -2.5127 -14.1256  0.8637
#>  -1.2591 -3.5274  7.5372  4.7249  7.0364
#>   4.2931  4.9043  2.6612  1.6353 -4.1055
#> 
#> (1,4,.,.) = 
#>   0.5959 -2.3818  0.1499  2.3706  2.7854
#>  -0.3407 -0.0652 -1.8789 -3.4756 -4.2882
#>   3.1855  3.5750 -3.0489  5.0809 -6.5056
#>   1.5921  0.3372 -6.8225  5.9676 -7.2015
#>   3.6415 -4.2509 -1.8287  1.5107 -4.4656
#> 
#> (1,5,.,.) = 
#>   0.4485 -4.7364 -3.7431 -2.6635  1.9138
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]