Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   3.8917 -0.7251 -0.2168 -1.5907 -0.8821
#>  -3.3615 -4.8495  3.9408 -6.0802 -0.2812
#>   3.2022  2.3691 -1.6002 -7.3786  1.0296
#>  -3.9963 -1.8347 -8.1132  9.9603  3.8300
#>  -0.4814 -8.3811 -0.5316  5.5540 -0.9876
#> 
#> (1,2,.,.) = 
#>  -0.0466 -0.3356 -1.7965  1.4569 -1.3515
#>   5.9369 -7.9302  8.6726 -4.6669 -10.1327
#>   2.9262 -1.4504 -1.7615 -8.9682  2.8003
#>   0.8736 -6.8159 -4.9975  6.4514  3.0176
#>  -5.9406 -10.8554 -4.8838 -1.2189  4.6013
#> 
#> (1,3,.,.) = 
#>   6.6626 -1.2171 -0.9496 -3.6591 -2.3004
#>   6.3784 -2.2955 -0.8252  3.3419 -2.0650
#>  -0.6958 -6.5582 -9.1582 -5.1316 -0.6263
#>  -8.2708 -11.7936 -8.2502  1.8270  4.7210
#>   3.0222  1.9115 -4.7272 -1.4235  1.6511
#> 
#> (1,4,.,.) = 
#>   -1.4512  -4.4162  -2.8669   0.5544   1.2592
#>    3.4418   3.9762   1.6510  -0.5103   0.0342
#>   -5.8698  -4.7206   6.7539   3.8798   7.5726
#>   -1.0872  -1.2868   1.7186   4.1220   4.2055
#>    0.5466   0.2296  -1.4232  13.1112  -0.3349
#> 
#> (1,5,.,.) = 
#>   -0.3144   3.8755   2.7112   0.5305   2.7114
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]