Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    4.6177   7.8996  -5.2030   4.5953  -5.8738
#>    2.4212  -2.0793  -2.4199   2.0154   4.1353
#>   -0.9072  -7.6279  14.2855 -18.7178  -7.1257
#>   -0.6928   1.0273  -9.1244  -0.6497   5.0983
#>    4.4894  -0.1908   1.2924  -0.8222   1.2195
#> 
#> (1,2,.,.) = 
#>    4.7412  10.3048  -3.8313  -0.5382  -4.6245
#>   -5.2846   8.6351   6.4429   9.5500   4.9874
#>    4.7743  -6.4466  -6.2698 -19.2720   5.2408
#>    2.6167  -1.0748   8.0501   0.5432  -9.5406
#>   -6.2633  -3.6368  -6.7502  -1.1890   0.7834
#> 
#> (1,3,.,.) = 
#>   -2.1816   8.5579  -0.8920   0.4865   1.1847
#>   -0.8923  15.6051 -14.7151  -5.5219   3.4599
#>    6.7935  14.3823  -5.0454  -9.1401  -1.1111
#>   -0.3399   4.9860   9.6702  -6.2324   1.0604
#>    0.2697   4.7540   1.9660  -7.4571  -7.7514
#> 
#> (1,4,.,.) = 
#>    0.2523   0.0801   3.4041   9.6708   4.7463
#>    0.3916  -8.9223  -6.4976  11.2998  -0.7003
#>    8.9817   3.5088 -10.5704  -8.6929   3.6959
#>    0.0180  -8.7898 -13.7742   0.6623  -4.8612
#>   -2.7607   6.2158  -2.9362  -0.2159  11.8706
#> 
#> (1,5,.,.) = 
#>   -1.3411   5.4679  -1.5871  -4.2223  -2.3837
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]