Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    2.2921   3.4949   3.7825   5.8963  -0.7031
#>    1.5317  -4.4133  -3.3712   0.7949  -8.1896
#>    0.4789  12.5294  -5.1576   0.7802  -3.5340
#>    1.1737  -2.5040   4.8033   9.6205 -10.3064
#>    1.9721   1.0463  -2.1732  -0.2967  -2.9611
#> 
#> (1,2,.,.) = 
#>  -0.5071 -5.0974 -2.6240  1.3640  1.6900
#>  -3.0640  3.3973  3.1869  4.7266  3.2430
#>  -0.8513 -10.0329  0.3850  0.0877  3.8624
#>   1.4610  8.4505  2.3699 -0.3091 -5.0730
#>  -2.5420 -6.4419 -0.7375  2.9278  3.7117
#> 
#> (1,3,.,.) = 
#>   -1.8883  -5.3352   1.9607   3.4500  -1.0135
#>    3.9117  -7.5179  -6.6964   6.8048   3.0023
#>   -3.8880   2.0819  -4.0249 -18.0830   3.1048
#>   -0.9999  -5.8814  -2.1756  13.9957  -1.8499
#>   -5.5848  -5.2947   3.8750 -11.7950  -2.0926
#> 
#> (1,4,.,.) = 
#>   -0.7559  -6.5390   2.0474   4.1379  -4.0667
#>   -0.4018  -1.0257  10.7388   8.5807   2.4472
#>    1.9261   0.8110   0.6613   3.1820   4.6845
#>   -4.3085   7.8791  11.8883  -5.3429  -5.1401
#>   -2.3737  -8.3453   5.5763   0.4831  -4.8141
#> 
#> (1,5,.,.) = 
#>  -5.2320 -0.2389  6.6026 -1.7347 -0.3692
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]