Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -0.9705   1.3454   3.5472   2.0369   1.0047
#>   -4.3948   3.1280 -11.4111  12.3706  -2.2507
#>    0.4358  -8.4513   2.0198  -2.3195   1.6715
#>   -6.3952   0.4049   1.4754   4.4663   0.7854
#>    1.8996   0.9878  -6.3684  -0.7751   3.1238
#> 
#> (1,2,.,.) = 
#>   3.7612  1.2743 -2.0969  6.4575 -3.5994
#>  -1.0571  2.9648 -0.7024  2.5732 -2.4886
#>   0.7687 -2.0233  6.5034 -3.9527 -5.0138
#>  -5.1076  7.6778 -7.3579 -0.6021 -3.3051
#>  -2.6213 -6.3518 -6.8017  5.2741 -2.3735
#> 
#> (1,3,.,.) = 
#>   -0.8496   3.5398  13.7673  -7.9973   0.4542
#>   -0.8893   3.0198   0.8213  -1.5904  -3.3611
#>   -4.1627  12.1782  -1.5519   8.5268  -3.4804
#>   -4.9486  -5.6361  -3.9393  -4.8518   0.9603
#>   -1.5680  -2.4794  -0.0853  -6.6395  -0.9866
#> 
#> (1,4,.,.) = 
#>   -1.3988  -5.2536   0.6466  -0.3379  -0.5035
#>   -4.2823 -11.4858   0.1428   0.8623  -7.8721
#>    0.8780  10.3979   2.0596  -4.7824  -3.0318
#>   -0.0768   6.9422  -5.6651   5.5460   3.8756
#>    4.6836   5.5778  -0.0347   0.2440   1.9103
#> 
#> (1,5,.,.) = 
#>   -0.1051  -2.2898  -0.8515   5.9179  -7.0499
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]