Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -0.3441   0.5670   3.0733 -11.2769   0.5115
#>    2.6821  -4.9539  14.5088   2.0780  -0.0160
#>    0.1589  -6.7839   5.5799   4.5381  -2.3080
#>    4.1413  -3.2047  -0.8855  -9.6738   3.4029
#>    2.1240   0.6366   1.9537 -12.4547   2.7049
#> 
#> (1,2,.,.) = 
#>   -2.1797  -3.1348  12.8171   4.8633  -2.1815
#>   -8.4042   3.9963   6.3934  -0.2003  -0.2018
#>   -0.6245   2.8818  13.6172  -7.8163  -6.8928
#>   -0.6966  13.5284   6.1512  -0.4216  -1.0590
#>   -6.5398   5.7357  -2.0816  -0.6272  -0.0087
#> 
#> (1,3,.,.) = 
#>  -5.6243 -3.3340  4.3450 -2.7043  1.5315
#>  -5.7892  7.2354 -9.2707  2.5890 -3.1599
#>   1.1701  4.8359 -8.9655  0.8060  9.0110
#>  -5.3337  4.2272 -9.8124  3.6656  7.8268
#>   2.5647  7.7334 -13.0517  7.6591 -5.4978
#> 
#> (1,4,.,.) = 
#>    1.6648   4.7733   0.4875   5.1230  -8.6828
#>   -4.2515  -1.8087  -2.6517   1.8086  -1.2615
#>    4.8142   9.4244   3.6554  -7.2004   0.1725
#>  -10.5056  12.0332 -14.8739  -0.7431  -0.1441
#>    4.6355   4.7776  -6.7722  -0.2721  -4.1180
#> 
#> (1,5,.,.) = 
#>   -1.0520   4.0885 -11.2460   2.0825   6.6329
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]