Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   11.4511   5.4423  -6.0107   1.1300   3.8783
#>  -11.1548   6.6599   8.8971  -0.6624   4.9520
#>    7.4389   2.3154  -3.6736   0.3223  -6.2985
#>   -9.0315  -4.7300  -7.8184   4.7333  -1.1794
#>    4.0458   6.2457  -2.1936   3.5356   9.5120
#> 
#> (1,2,.,.) = 
#>  -8.1533  2.8866  2.9244  2.7152  0.8692
#>  -2.7728 -4.4609 -5.5061  1.6426 -3.3097
#>   6.8238 -3.9615 -2.2761  0.5819  4.8390
#>   3.2162 -2.7878  0.5012 -0.5397 -4.0439
#>   5.8829  7.1273  7.5877 -9.0971 -3.0380
#> 
#> (1,3,.,.) = 
#>    4.7176   2.0863  -2.5185  -2.4536  -1.4936
#>    4.0063  -5.2302  -4.0595  -4.9428 -10.3370
#>   -7.6836  -7.0003  -0.5365  -5.0301   5.3008
#>   -1.5480  -0.5785  14.4789   2.0025   8.9555
#>    7.8354  11.0405   5.5523   6.8135   7.1549
#> 
#> (1,4,.,.) = 
#>   12.3693  -2.6110  -1.8462  -2.7681  -0.1439
#>   -8.2491   9.0108  -2.4620  -2.6971   6.0339
#>   -1.8999   3.4420   5.0806   6.3226   0.5239
#>   -0.2717  -1.1447   8.8898   3.7146   2.6570
#>    3.7523  -3.4554   2.4549   4.9909   9.6051
#> 
#> (1,5,.,.) = 
#>    1.5977   1.4040   4.2622   0.3781  -6.9362
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]