Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    5.7629   9.0288  -1.1745  -2.8176   0.6222
#>   -7.4638  -1.7343   1.3503   3.2497  -1.5780
#>    4.4125  -1.4259  -5.9147  -4.7154   0.2897
#>    0.3425   0.6058  -0.0461   1.0502   2.7429
#>   -6.6821   3.7508  10.6363  -1.0241  -3.5343
#> 
#> (1,2,.,.) = 
#>   4.7855  7.2238  4.3464  7.1195  5.0932
#>  -8.0387 -4.2870 -17.1267 -2.9716 -1.2198
#>  -3.8417  1.8821 -0.9253 -4.5654 -5.0634
#>   5.4444  4.9778  5.0004  4.2699  2.4388
#>   4.5790  2.7188  2.3895  1.6850  1.0489
#> 
#> (1,3,.,.) = 
#>   -1.0172  -5.5906  -9.2500   1.1740   3.4473
#>   -6.7668   7.7105  16.0405   3.8227   4.3213
#>    8.1370   6.9322   4.3332   3.4332   0.3311
#>    5.5541   2.4267 -13.6225  -3.6443  -2.8270
#>   -1.5993   2.3097  -0.6151  -5.0470  -2.5652
#> 
#> (1,4,.,.) = 
#>   -9.5844 -11.6151   1.9417   1.4473  -0.6089
#>    1.2294  12.4144   5.0775   6.3693   1.7599
#>    5.3715  11.5626  -1.5411  -6.2147   4.5789
#>   -2.5553  -1.4738  -4.0765  -0.9799  -7.1599
#>   -8.8361   2.5553  -8.3771   7.9612  -3.7150
#> 
#> (1,5,.,.) = 
#>  -0.2209 -3.8245  5.5160 -0.8640  2.2418
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]