Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -3.7977  0.2829 -0.9268 -8.2560 -4.7614
#>   0.1400 -4.8221  6.2701  4.8074  4.7207
#>   1.0233 -2.3552 -6.6120  8.8624  6.9712
#>  -3.4671 -6.4584 -1.6016 -6.2723  2.1300
#>  -2.7220 -8.6724 -7.7918 -2.5265 -5.7396
#> 
#> (1,2,.,.) = 
#>    5.1352 -11.6039   0.9371   6.9367  -5.8340
#>   -1.3139  10.7532   1.7803   4.3282  -0.8915
#>  -11.8359  14.6625  -4.1376  -0.1215   4.7926
#>   -6.3065  -3.5012  -6.8489   3.8888  -2.0203
#>    6.6240   1.3412  -4.2936  -2.4049  -4.7689
#> 
#> (1,3,.,.) = 
#>    6.6035   4.4499  -4.4834  -7.0443  -4.0309
#>    0.0575  -0.6161  -5.6741   5.3717   2.3695
#>   -3.8027   8.7036   0.2339   7.0629   2.5125
#>   -8.2105   0.0771  -4.0971   3.2980  -7.8003
#>   14.1254  -1.9736  -9.7468  -3.6884  -2.2229
#> 
#> (1,4,.,.) = 
#>   -1.0505  -6.7482  -1.1963   4.3107   7.4720
#>   -3.2113   4.3933  -0.6698   9.4009  -6.5958
#>   -4.0358  10.1992  -2.8802  -1.2376   4.5386
#>    7.6166  -8.1955  -1.2387   1.3450   1.3544
#>    0.9911  -9.1904   0.9460  -6.1591   2.2151
#> 
#> (1,5,.,.) = 
#>  -1.4196 -6.2667 -0.9944  0.1546  5.4247
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]