Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -0.8924  -2.0889   1.1467  -3.3314   4.6627
#>   -5.0985  -0.6315   0.3915   8.0673  -0.2173
#>   -3.0950  16.3506  12.1716  -8.2273  -3.0338
#>   -8.8274  -5.2141  11.9851   4.4503  -6.5317
#>    2.4386  -1.1186  -0.8584   2.0805   1.6188
#> 
#> (1,2,.,.) = 
#>  -0.0532  1.8522 -3.1775  0.9585 -2.2285
#>  -3.5316 -2.7154  3.3351 -4.3313 -0.2427
#>  -5.5162 -7.8008 -1.0600  9.6093 -6.9194
#>   3.9603 -1.2625 -3.9162 -1.9209  6.4941
#>  -1.6873  1.4923 -2.4747  1.6502 -5.6107
#> 
#> (1,3,.,.) = 
#>    0.2040   1.0242  -3.1136  -7.8209   2.5712
#>   -2.7702   1.0033   3.4507  12.3343  -4.5161
#>   -3.1260   3.6548  13.6104   2.0188  11.3534
#>   -0.1965  -2.8416   1.9790   7.1586  -3.1812
#>    2.5831   3.7116  -3.1094   0.9260   0.7625
#> 
#> (1,4,.,.) = 
#>    3.1479   2.0931  -2.0171   3.5739   3.1002
#>   -4.2053  -2.1371  -3.3433  15.3165  13.0872
#>    2.1652  -4.4607  -1.0221   4.0273   9.3409
#>    1.8631  -9.9568 -12.4754  -3.5444  -2.2463
#>    0.2693  -2.9323   0.9312   3.0957   1.9440
#> 
#> (1,5,.,.) = 
#>   1.1197 -4.5852 -10.3621  1.7538  4.9901
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]