Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -1.7853 -2.7672  4.6838 -2.1012 -3.6708
#>  -4.0970 -3.6665  2.1587  1.7290 -1.7238
#>   4.9711 -10.5130  3.4417 -1.6098  2.3305
#>   5.4525 -3.3414  1.1987 -3.9711 -2.5596
#>  -0.3002 -4.0342  2.1409 -1.3450 -4.1959
#> 
#> (1,2,.,.) = 
#>   0.2269 -7.4921 -6.2246  0.7198 -3.1205
#>   5.0485 -7.3670 -7.7470  4.6573  6.9922
#>   0.9512  2.3345 -7.0022  2.4359 -4.9371
#>  -0.2500  6.5136  0.9957 -4.6299 -4.6762
#>   3.8448  9.0486 -3.5823 -0.0050 -2.5861
#> 
#> (1,3,.,.) = 
#>   2.8031 -9.9173 -10.6770  0.5296  4.9224
#>   8.1257 -0.7097  6.8791  4.5384 -1.0124
#>  -5.6121  0.7324 -5.5166 -3.9524 -5.6021
#>   4.6058 -11.4215  1.0963  1.3372 -3.9481
#>  -1.8657  5.5494  3.9547 -4.9163  1.9631
#> 
#> (1,4,.,.) = 
#>   1.5932  4.4284 -5.8843 -7.4152 -0.0763
#>   3.8571  7.7701 -2.4589  9.5240  8.9927
#>   1.6958 -9.6852 -3.1822  7.6399 -15.1873
#>   2.3516 -3.8421 -5.9734  6.4797 -8.1333
#>   0.5821 -10.8696  1.2234 -0.3817 -3.1971
#> 
#> (1,5,.,.) = 
#>    1.8972  -6.6898  -6.9244   3.1374  -3.9452
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]