Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -0.7317  6.5211 -1.5028  3.8357 -2.4103
#>   9.8186  1.5393 -5.2660  4.6897 -4.0131
#>  -0.9060 -0.1176  4.7689 -1.6379  0.3057
#>  -2.1466 -3.9723  3.4982  2.7933  7.4747
#>  -2.7505  0.3500  2.0932  7.7128  2.6084
#> 
#> (1,2,.,.) = 
#>   -5.6564  -5.6607   0.0554 -10.6488   1.0495
#>   -1.0315  -3.4584   6.8347  -0.1561   4.1761
#>   -2.4862   4.3325   7.1202  13.1530  -5.3877
#>   -2.8189 -14.2718  -5.3121  -2.6760  -7.4943
#>    3.9376   4.5215  -0.8428  -9.3841 -12.0134
#> 
#> (1,3,.,.) = 
#>   -5.8602   2.7079  -8.7718  -1.3515   5.1433
#>    1.9664  -1.0484  -5.4223  14.1826   2.7572
#>    0.2881   1.0240   2.7793  -5.1067   4.4854
#>    4.8752   3.2712   2.9220  -5.9801   0.4432
#>    1.3796   2.5954  -1.4176  -5.2357  -0.4691
#> 
#> (1,4,.,.) = 
#>   2.6331  0.3881  0.2792 -4.1988 -1.6097
#>   3.8176  4.8020 -1.7139 -4.8141 -2.1381
#>  -1.5298 -5.9795  6.9098 -5.5665 -4.9812
#>  -4.6422  4.6957  5.8748  6.4625 -8.7214
#>  -2.0847  0.2269  1.4184  7.3928 -1.9048
#> 
#> (1,5,.,.) = 
#>   2.8963 -3.3789  1.8631  0.1953  4.9935
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]