Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -0.2600   1.5780  -0.4636   5.2538  -1.9254
#>    8.8889  -3.1290  -0.6187   0.0917  -3.8467
#>   -1.6645   3.6187  23.8119  -1.4101  -4.5648
#>    0.8994   3.6648  -1.2102 -17.2275   1.0792
#>   -1.0140  -1.2454  -4.3946  -2.7393   3.2143
#> 
#> (1,2,.,.) = 
#>    0.5519  -1.1728   5.8941  -3.6558   1.3719
#>   -4.9662   4.0351   5.4949   3.6160   2.9684
#>    2.6367   9.8184 -12.4837   7.4180  -3.0289
#>   -1.2171  10.5992   7.6100   2.7905  -6.5756
#>    3.6078   4.6638   5.0421   1.8211   1.4273
#> 
#> (1,3,.,.) = 
#>  -3.8579 -8.1489  1.2833 -3.6431 -1.8809
#>  -3.1395  3.5042 -2.6050  3.5646 -0.9589
#>  -0.7295 -7.2721 -6.5696 -4.4387 -0.3718
#>  -1.4046  5.0623  5.2375  0.6787 -1.9557
#>   4.3407  7.7225  2.0663 -2.3769  8.0254
#> 
#> (1,4,.,.) = 
#>   -0.1847  -6.6731  -0.3476  -1.6306   0.7671
#>   -7.9129 -15.2408  -6.6462   3.5566  -0.0110
#>    2.2209  -0.3275   3.1960   4.3015   2.1565
#>   -0.1236   2.0273   5.2620  14.1095   8.4341
#>   -3.5959   1.3574   3.9114   7.4811   4.7685
#> 
#> (1,5,.,.) = 
#>   -9.6705   1.9456  -6.8314  -0.4936  -1.9184
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]