Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   2.9245  7.0243  4.1803 -7.8310  1.9749
#>   6.9954 -4.5371  2.6155 -0.6258 -8.5163
#>  -6.4761 -3.6470 -4.6900 -2.0254 -1.0044
#>   1.7357  8.5697  5.0527  0.5047  2.1309
#>   4.1236 -5.6353  0.4639 -6.0977  4.1048
#> 
#> (1,2,.,.) = 
#>   -8.2067   0.5444  -1.1571  -0.9292   0.9086
#>    1.0599  -6.0254  -8.6750  -3.2027  -3.4542
#>    6.3506  -0.4982  -0.9502  -4.2749   4.0699
#>   -1.7904  -7.3960  12.1528  -6.0314   6.4512
#>   -5.3375  -0.2332  -3.3655  -3.1853   1.6503
#> 
#> (1,3,.,.) = 
#>  -1.7366 -3.1552 -3.3595 -3.3799 -1.0171
#>   4.7646 -1.0476 -3.6258  3.1142 -5.2829
#>   2.0112 -2.3323 -5.3359 -10.2330  2.4431
#>  -0.4641  7.4022 -2.9593  1.4940  7.0093
#>  -1.9704  8.4718  2.2171  1.3135 -1.3043
#> 
#> (1,4,.,.) = 
#>   -1.9741  -3.6854  -5.1384  -5.6632  -0.7166
#>    1.5365   9.8956  -1.7264  -7.2178  -9.8660
#>    2.6335  -2.1986   5.4083   1.5582  -2.6776
#>    2.4762  12.6252   1.8879   0.1153  -5.4631
#>   -1.2608   4.9088  -4.7999   6.1481   0.5553
#> 
#> (1,5,.,.) = 
#>   2.1926  5.9836  2.7723  4.8861 -3.9196
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]