Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    8.4713 -12.1481  -3.1618   4.4814  -0.6857
#>    3.2007  -4.9777   0.8580   4.6206  11.5135
#>   -6.1504  -4.1299   5.2561   3.7383   2.1906
#>   -3.7908  -0.2639   8.1627  -8.1484  -3.9120
#>    2.6081   2.8610   0.5994   1.6804   4.4392
#> 
#> (1,2,.,.) = 
#>   4.3801 -1.3456  0.9181 -0.4410 -6.2503
#>   0.2142 -1.5869 -3.7349  2.1712  0.3964
#>  -5.3835  4.6372 -3.5600 -0.9168 -2.1227
#>   2.3626  3.8383 -5.0942  1.0822 -2.4297
#>  -3.6757 -2.1757 -0.1075 -0.7089  3.3495
#> 
#> (1,3,.,.) = 
#>   10.1189   1.4809  -4.3071   4.1558   4.7755
#>    2.6659   6.7512  -0.0219  -3.9975  -2.9179
#>   -5.5147  -0.0629   4.4786  -3.6803   0.5827
#>   -2.6255  -8.1551  -3.0018  -3.7409   4.4801
#>    2.1670  -2.1770   0.2542  -9.5194  -9.3290
#> 
#> (1,4,.,.) = 
#>   11.6292  -2.7057  -2.4016   2.0065   0.4323
#>    4.4376  -8.7412   7.3867   2.9152  -0.2721
#>    4.9096  -5.1525   3.1420  -2.6245   4.2719
#>    1.9205  -8.2757   7.9558   2.4859   1.1324
#>    3.5174  -2.8704   2.3029  -2.6928   0.3976
#> 
#> (1,5,.,.) = 
#>   3.1021 -2.9272 -2.0440 -10.7058 -1.5633
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]