Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    5.7140  -4.7425  -4.5703  -2.5408   8.0833
#>    0.9545 -23.1909   2.7695  -8.1813   2.9908
#>    4.7993   4.6184   9.0967  -7.7689  14.0784
#>    1.2856 -16.2820  -5.5323   1.8652   1.3754
#>   -3.9915 -12.5675  -2.6568  -3.7553  -2.1214
#> 
#> (1,2,.,.) = 
#>    4.5618  -1.5330  -2.0375   6.1811  -1.6993
#>    2.9000   2.8434  13.5006  -4.2293  -7.1208
#>   -5.1473   1.9047  -8.9976  -3.6730   7.1624
#>    3.4233  -9.4147 -12.7942  -1.4240  -1.0182
#>    6.3534  -4.0970  11.2559   6.9799  -0.3569
#> 
#> (1,3,.,.) = 
#>    0.2242  -5.9848  -1.8764   0.4806  -1.9548
#>    8.7144  12.5933  10.2030   7.8701  -1.4370
#>   -1.0269  -0.8898   0.5095   0.5482   3.6305
#>    0.0801  -6.2994   3.2761  -1.4682  -8.1079
#>   -0.2233  -3.1006   8.1166  -5.8568   7.6171
#> 
#> (1,4,.,.) = 
#>   1.6276 -0.3334 -2.1776 -1.9248  3.8563
#>   2.0993 -7.5253  6.8891  4.9680 -6.0126
#>   0.0054  1.0899  3.7617  3.7532 -2.2708
#>  -5.1287 -7.1365 -3.6704 -5.5157 -6.1248
#>   0.2389 -10.3315 -1.8772 -12.9130  3.8851
#> 
#> (1,5,.,.) = 
#>    0.3130   5.4533  -3.0663  -0.7376  -2.5326
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]