Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   3.7498 -3.1155  0.5420  0.9744 -5.9905
#>  -5.0210 -2.6714  6.5989  2.1032  0.7233
#>   4.1044  7.8272  4.4387  1.0071 -1.3675
#>   2.3901  1.0210 -3.1322 -1.8445 -5.7807
#>  -1.7530 -3.2944 -5.0934  2.2690  3.1555
#> 
#> (1,2,.,.) = 
#>   2.0266 -2.4575  1.5559  1.6819  0.5363
#>   4.3391  5.7472  2.4879 -5.4799  6.0574
#>   3.7523  0.5048 -0.5972 -0.4957 -5.4948
#>   6.6094  2.4788 -0.1453  4.0563  4.0902
#>  -0.3984  5.0399  6.4420  7.5586  3.6895
#> 
#> (1,3,.,.) = 
#>   1.3100  5.3118  3.7760  0.9806  0.4565
#>   2.3241 -2.0477  5.0608  1.5986 -1.7347
#>  -1.0433 -1.9256 -1.7242 -2.6099 -6.8180
#>  -0.7126 -1.3488 -1.3288 -1.3871  2.7182
#>  -1.1309  6.5699  1.9839  7.3877  0.1692
#> 
#> (1,4,.,.) = 
#>   3.3187  1.7995 -3.1301 -0.1027  3.3633
#>  -0.6603 -0.0096 -7.3141 -3.5015 -0.5290
#>   4.9382  4.5153  0.2664  0.3836 -2.3189
#>   2.9798  2.3563 -2.9647  2.0562  3.7435
#>  -0.6254 -4.0056  4.7182  1.9440 -0.7034
#> 
#> (1,5,.,.) = 
#>   1.3556  1.3716  4.7668  5.5047 -3.1514
#>  -2.0814 -0.2609 -4.7637 -6.4941  2.3308
#>   0.7484 -5.6539 -10.3683 -0.3031 -5.0198
#>   2.3921 -7.4785  0.9051  5.7889  7.1976
#>   8.3061  2.0583 -1.3598  2.0524 -5.4920
#> 
#> (1,6,.,.) = 
#>  -1.0835 -0.9244  3.5873 -0.5053  3.7351
#>  -1.9874 -5.8592 -2.1367  3.5609  1.7312
#>  -5.1390 -2.2748  6.5253 -9.0355 -6.2111
#>   4.0062 -4.7559 -2.3554  9.5085  2.3048
#>  -2.4125 -1.8464  4.1743  2.9814 -1.1094
#> 
#> (1,7,.,.) = 
#>    4.5502   2.6221  -1.3769   2.3807  -6.8988
#>   -4.3726   6.1141   2.5510 -10.2892  -4.6505
#>    0.3608   4.9268  -6.2002   1.4095  -2.1463
#>    4.0118   3.9910  11.0088   9.5550   2.8846
#>   -3.3122  -1.6118   3.2318  -1.4398  -1.9942
#> 
#> (1,8,.,.) = 
#>   4.9177  0.5940  0.9629 -1.7863  4.9310
#>   1.9738  3.8593 -2.8888 -11.8238 -3.7437
#>  -0.5077 -0.1873 -1.4976  0.4488 -0.2435
#>   1.9120  9.7473 -4.4501 -4.8311  8.7993
#>   0.7416  5.6446  4.8415  6.6458  0.9768
#> [ CPUFloatType{1,8,5,5} ]