Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    5.2489  -0.6952   7.4523   2.7252  -2.1083
#>   -1.8294   0.0461   1.6267  14.0682   1.9279
#>   -2.7049  -4.5564  -0.9723   7.0989  -1.2964
#>   -2.8631  -9.1348  -1.3644   8.7801   8.5420
#>   -4.3145  -3.7297   1.5278  -6.0266  -0.6739
#> 
#> (1,2,.,.) = 
#>  -0.1821  3.0858  5.4415 -8.0598  1.2184
#>   5.9714  9.9285 -19.3598 -2.2519  2.3365
#>  -7.6735 -9.3232  7.1277  5.0188 -7.7350
#>  -0.0565 -1.8478  7.0045  4.5682 -3.0342
#>   2.7053  6.4341 -8.2598 -3.8015  1.8236
#> 
#> (1,3,.,.) = 
#>    8.5392   1.6636 -10.4516   6.0930   3.4160
#>   -4.6261   4.5839  14.1632  -0.2775   0.5904
#>    7.1027  -3.7751   2.8910   0.1585  -0.3573
#>   -7.9336  -2.8281   2.2523   1.2308   6.0588
#>    0.3408  -7.1548   2.1019   5.5082   4.4558
#> 
#> (1,4,.,.) = 
#>    2.0991  -1.6890  -4.4557   3.4598   3.5535
#>    7.9755   8.5312   1.1122  -1.6995  -5.7229
#>    4.3500  -1.5901  -1.9324  15.7307  -3.7528
#>  -10.2392  -1.7565   0.1876  -0.2210  -3.7637
#>    0.2663   0.9670  -8.7296   2.3473  14.5738
#> 
#> (1,5,.,.) = 
#>    1.2110  -3.1281   2.4735   4.2393   0.9490
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]