Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -1.6472   1.8027   1.0605  -5.7660  -2.7619
#>   -0.4898  12.0691  -2.6366  -6.2166  -3.0038
#>    9.1250  -2.4213  -4.4790  -9.6749   0.7035
#>   -5.9300  -7.9550  -0.1836   6.7058   0.0356
#>    2.2391   7.3690  -3.3928  -5.3290   0.6037
#> 
#> (1,2,.,.) = 
#>   0.2576 -0.0402  7.4012  3.1539 -1.8953
#>   0.8823  0.6829 -3.2528  4.8784  1.9237
#>   0.5158  1.2632  7.8948  0.8041 -6.6355
#>  -2.1970  6.3249 -0.6384 -0.9458 -2.8339
#>  -3.1609 -2.8035 -4.0993 -0.2023 -0.4497
#> 
#> (1,3,.,.) = 
#>   -2.9679   1.0516  -0.8488  -4.7248   2.3609
#>   -3.1756   3.4750  -6.0795   2.3277   4.9862
#>   -2.0545   1.1618  -2.8518   3.5332   6.4680
#>    5.5728   0.1598   5.5101  10.2637   3.1790
#>    1.2574  -0.0564   3.4414   9.2196   0.8992
#> 
#> (1,4,.,.) = 
#>  -0.6883 -0.1550 -2.8211  1.4332  0.2834
#>  -2.0089 -1.3660  3.5141 -0.9596  4.6634
#>  -2.9187  0.6930  3.1485  5.0158  2.1943
#>   3.3887  4.1258 -0.5359 -5.0693  0.1015
#>   2.8523 -3.9480 -2.4553  2.1379 -1.2994
#> 
#> (1,5,.,.) = 
#>  -1.1318 -4.3666 -2.4024 -0.8313  2.8860
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]