Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    0.2412   2.1005   6.6234  -5.5929   4.9340
#>    2.1103  10.2606  -5.3042   4.0971   1.7491
#>    3.3527  -1.8740   2.5444  -0.7046   1.4613
#>   -2.7335   0.3506  -1.6185  -0.8445   3.2116
#>    2.3850  -7.1242   1.7694  -0.8270   1.7975
#> 
#> (1,2,.,.) = 
#>  -2.4107 -2.0328  7.0021  3.5637 -2.8205
#>   6.4856 -1.7348 -1.7044  1.4380 -0.1942
#>   0.1709  6.5373 -7.5761  2.3001  2.7411
#>   0.0724 -8.1260 -2.7169 -0.6458  3.1061
#>  -1.5382 -3.7197 -2.6390  0.2539 -2.3014
#> 
#> (1,3,.,.) = 
#>    0.9703  -8.7512  -1.6888  -6.0769   0.1415
#>   -0.3903   8.4972  -3.9673   6.2363   1.6247
#>    2.3259   0.1894  -8.3374   3.0667  -7.1511
#>   -5.9655   9.6314  -2.0416   2.3261   2.9605
#>    4.6337  11.8126   2.4667  -0.2484   2.0264
#> 
#> (1,4,.,.) = 
#>   -6.5037   2.1178  -4.4248  -1.9224   4.2397
#>    6.9817  -6.2822  11.9385   0.9831  -4.8272
#>   -1.8688  11.7422  -7.1046  -3.4722  -0.4431
#>   -1.4049  -3.1978  -6.0436   2.7018   2.7226
#>   -1.4194  -4.1370   5.0302   2.9417   3.2194
#> 
#> (1,5,.,.) = 
#>    1.1275   0.2527  10.0121  -8.9137   6.1655
#>    3.9311  -5.7285   0.7211  -1.0921   4.9324
#>    4.8780  -6.0988  -1.5595   0.2154  -2.2197
#>   -5.4994   2.4119  -3.5367   4.4083  -3.9627
#>    1.6790  -4.0881  -1.5034  -3.4489   6.8829
#> 
#> (1,6,.,.) = 
#>  -1.0160  3.3103  1.0483  2.6920 -1.5154
#>  -3.0043 -1.7246 -1.8580 -0.6739 -2.9448
#>   1.2264  3.0546 -1.3289  5.4251 -0.7095
#>  -5.6945 -1.1924 -4.9008  1.6715  2.6690
#>  -3.8813 -4.7026 -3.4411  4.8084 -3.0748
#> 
#> (1,7,.,.) = 
#>  -1.6594  8.7352  0.3221 -1.9164  0.1215
#>   8.3073 -13.2826 -1.2722  0.8839  2.6828
#>  -2.0230 -1.3708 -1.5325  8.7706 -1.8196
#>  -2.8281 -3.2151  1.1348 -1.7700  0.3153
#>  -3.6870 -4.1955  4.9134  7.3294 -0.9155
#> 
#> (1,8,.,.) = 
#>   -0.2824  -4.4791   2.1272  -4.3638  -5.1920
#>   -7.2475   2.2503   3.0449   1.8146  -1.1592
#>    3.4866  -3.5456   1.7442   0.5284  -3.7938
#>    0.6542  14.7206  -9.1515  -7.6679  -2.8497
#>    4.7160   2.9921  -7.5768  -4.1088   1.7886
#> [ CPUFloatType{1,8,5,5} ]