Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -2.9916 -0.2528 -4.6060 -1.8395  2.9395
#>  -0.5892  8.0865  3.8223 -0.8167 -3.6720
#>  -1.1749 -0.2046 -0.2680  6.4752 -2.4032
#>  -2.8153  1.7670 -1.4777  9.9320 -2.7266
#>  -2.2218 -1.4163 -3.3006  9.5650  0.1838
#> 
#> (1,2,.,.) = 
#>  -3.2249  1.0960  0.2183 -3.2967  2.6181
#>  -6.0432 -8.8603 -4.3476 -7.6009  1.5652
#>  -2.3138 -5.4033  1.8584 -1.3602 -0.9847
#>   2.7415  8.7225  1.3217  0.6317 -1.3156
#>   1.7976  4.0229  5.7739  3.6604  3.1012
#> 
#> (1,3,.,.) = 
#>  -5.9846 -3.0113 -11.7441 -3.2833  4.7583
#>  -5.0218 -1.9542  6.8070 -1.4147 -2.3353
#>   3.3824 -3.5654  6.5112  9.6738  1.5727
#>   7.0188 -2.7420  3.7557 -4.6396  5.2343
#>   6.9281 -6.6578 -4.1421 -2.1582  4.6370
#> 
#> (1,4,.,.) = 
#>    6.6410   3.9535   0.7665  -4.8741  -2.7023
#>    3.5687  -6.5623   3.2446  -2.6057   1.5128
#>    1.4579  -5.5561  16.3209   3.4935  -1.0701
#>    0.3309 -17.6475  11.9170   7.3619  -0.7922
#>    6.8925 -12.2061   0.0347  -0.8927   1.4056
#> 
#> (1,5,.,.) = 
#>  -2.5828 -4.6728 -8.6236 -7.5377 -1.4394
#>  -9.6101 -2.0141 -4.0620 -5.4692 -7.7653
#>  -2.8430 -4.9317 -4.2425  6.6001  7.6349
#>  -6.8661  1.7996 -4.1608  2.8609  1.6560
#>  -2.6439  0.0894 -9.7275  5.5682  2.0675
#> 
#> (1,6,.,.) = 
#>  -0.2229 -10.1986 -1.6046 -5.1201 -2.8190
#>   5.4422 -11.2753 -3.6540  9.1923  2.3467
#>  -1.1844 -4.8289  5.8708 -2.5023 -3.2680
#>   0.0937 -2.4135  6.0026  1.3198 -4.4941
#>   0.3714 -3.6307  5.4265  2.2516  1.4755
#> 
#> (1,7,.,.) = 
#>   -3.6251  10.6643   5.4242  -4.8898   5.1827
#>   -4.1256  -2.1213  -1.4159  -5.0265   3.0149
#>    4.5904   3.6101  -1.4680  -9.3641  -2.8581
#>   -2.4562   0.4594   2.5265  -2.9454 -13.5782
#>    5.1761  -7.9144   5.8999  -2.0799   1.1649
#> 
#> (1,8,.,.) = 
#>  -2.3171  3.7989  2.3058 -0.3894  2.8399
#>   1.9401  2.4085  5.7565 -3.3360 -2.8440
#>   3.9400 -9.3910  8.2221  6.1266 -3.3380
#>   1.2152 -3.5085  1.8401 -9.0912 -5.0809
#>   6.3660 -7.3800  6.9307 -2.0374  7.6020
#> [ CPUFloatType{1,8,5,5} ]