Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   -1.8473   4.0085  -9.6729  10.0332   3.7979
#>   12.2402  -1.8596  -7.9444  13.2653   3.5889
#>    4.0403  -1.2781   8.3182  -3.5255   9.2390
#>   -2.3504  20.0761  -4.4291  16.2662   7.0466
#>   11.2142  13.2551  -0.7120   5.4132  -2.9857
#> 
#> (1,2,.,.) = 
#>    5.2419   0.4859  -0.0363   0.1624   1.9499
#>    0.7757   8.7684  12.7201   3.8691  -1.2172
#>    0.3329   2.3769   0.6795  -4.6097  10.2930
#>    3.9475  11.6511   0.9870  -0.4857  -3.4191
#>   -7.3542  -1.3754   1.5724  -4.7907  -2.4778
#> 
#> (1,3,.,.) = 
#>   0.6623 -0.3360 -9.2527  1.5198 -5.3189
#>  -2.5656  3.0601 -10.2792  4.1255 -0.4105
#>   3.0901 -3.6973 -5.2128 -4.2379 -1.3200
#>   1.4448 -7.1420 -6.5785  2.6340 -8.7450
#>   2.5498 -7.9952 -5.5233  6.3409 -0.2703
#> 
#> (1,4,.,.) = 
#>  -6.4556  1.8909 -2.9857  3.2479 -5.4176
#>   7.5338  0.0531 -4.2485  0.8694 -2.0060
#>  -12.6371 -1.5396 -2.7723 -2.4450  4.2167
#>  -4.3191 -1.7908  0.4574 -0.4141  2.3007
#>   6.7356  9.2332  1.4251  3.1252  3.1259
#> 
#> (1,5,.,.) = 
#>  -0.9170  0.7915 -4.6154  9.0050  0.0433
#>  -2.4205 -3.3427 -2.6378 -8.2470 -4.2062
#>  -2.1171 -0.8815  2.7426 -10.4162  2.1091
#>  -5.4549 -6.0555 -6.6433 -7.0772 -8.9735
#>  -2.7601 -2.3348 -3.0432 -4.1442  7.1611
#> 
#> (1,6,.,.) = 
#>   -3.2584   2.5879  -2.9650   3.6556  -4.0091
#>    2.0103   9.7632   3.1749   3.1126  -0.6830
#>   -5.8191   6.6294   1.3383  -3.2929  -3.5561
#>    1.0747  -4.3152  11.3627  -5.9415   6.0667
#>    2.4337   6.5742  -4.8674   0.4513  -0.4388
#> 
#> (1,7,.,.) = 
#>   3.2495  0.6773  0.5127 -2.6367 -2.4910
#>   4.2821 -8.1398 -15.1244  3.9889 -6.5386
#>   2.9808 -3.8316  6.2569 -2.6304 -2.2474
#>   4.2669  3.2195 -9.6754  3.4172 -2.0675
#>   3.8295 -4.1143 -8.4488  1.5172  3.8538
#> 
#> (1,8,.,.) = 
#>    5.8813  -4.8008  -1.5713  -1.2545  -0.2361
#>    4.9106   9.0007   0.5698   9.7651  -3.5224
#>    1.1723  10.2527   3.6510   5.3991  -5.6769
#>    3.7163   8.4968   1.7147  -7.7119  11.0516
#>    3.0504   4.3694  -9.6801  -0.4388  -0.5715
#> [ CPUFloatType{1,8,5,5} ]