Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   1.0004  5.9655  6.0206 -1.7866 -4.0497
#>  -2.4533  8.4905  6.0862  1.1922  0.0739
#>  -6.0406 -4.9878  2.3122 -4.9762 -1.4355
#>  -3.4214 -6.2434  2.1530 -0.3671 -2.0343
#>  -1.9115 -6.9763  7.0927  5.8276 -4.2201
#> 
#> (1,2,.,.) = 
#>    0.7679  -3.7260  -4.7167  -1.9082  11.1248
#>    1.1792   0.5685  -6.8577  -5.5148   9.4854
#>    8.9220   1.2404 -13.0393  -3.4527   6.5251
#>    6.7142   1.0776  -5.3968 -10.1345   0.5096
#>    0.0472  -0.1316  -4.0639  -9.1367  -1.8133
#> 
#> (1,3,.,.) = 
#>  -0.1488  4.7068  4.9812  3.0144 -0.0452
#>   2.0136 -0.8656  0.6678  2.6514 -4.6636
#>  -4.6446 -6.2411 -6.2666 -4.4910 -0.8789
#>  -0.2734  0.7267  6.6415 -7.2319 -4.8246
#>   0.8521  0.1069 -0.9014 -1.7170 -1.0815
#> 
#> (1,4,.,.) = 
#>   -1.7400   2.6460   0.2398  -2.9762  -2.6205
#>   -5.7554  -2.6427  -3.7748  -0.7155  -0.0119
#>    3.7553   3.5842   3.8075  -1.3794   1.0097
#>   11.8116  12.1985  -2.3949  -0.7103  -2.0308
#>   -3.3760  -0.5621   0.2302   0.9772  -1.1270
#> 
#> (1,5,.,.) = 
#>  -3.4117  2.0148  8.0057  6.1525  1.9208
#>   1.9803  1.3856  0.2969  4.2163  3.5894
#>  -6.7136 -9.4852  2.4792 -2.2889  0.2243
#>   1.2009 -5.7582 -5.0125 -0.9677 -2.2046
#>   5.2396  1.0995 -3.6800  2.3566  4.0598
#> 
#> (1,6,.,.) = 
#>   -2.9629  -6.2210  -2.8443 -12.2926  -1.1794
#>    5.3607   9.0591   4.5425  10.4000  -8.2504
#>   -1.4764  -0.5863  -4.4092  -2.9450   2.7482
#>    1.9249  -1.8265 -15.6066 -11.3898  -7.7305
#>    2.2010  -1.0543   7.8164  12.0819   5.7975
#> 
#> (1,7,.,.) = 
#>  -2.7923  0.3437  5.6212 -5.9191 -7.6633
#>  -3.6437  5.9869  8.6759 -1.1109 -0.0193
#>  -0.7828  1.3387 -1.2718 -0.3724 -0.8216
#>   1.3469  4.0320 -5.7575 -1.9155 -5.6008
#>  -3.4712 -5.6221 -8.7107  5.0469  0.7825
#> 
#> (1,8,.,.) = 
#>   3.8398 -4.2646 -3.9171  5.0469  4.6974
#>   1.5134 -3.5499 -2.1115  2.1013  0.7474
#>   1.1365 -1.3102 -6.0034  5.4332  8.5716
#>   5.6955 -0.1683 -5.1332 -4.8705  7.6884
#>   0.4025  1.7707  0.8850 -4.4062  8.5310
#> [ CPUFloatType{1,8,5,5} ]