Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -3.4331  1.1846 -6.1142  3.7535 -4.2951
#>  -2.9584  2.9409  7.3701 -5.2912 -5.0721
#>  -5.8043  2.8642  3.4534 -9.5723 -1.4370
#>  -3.8673 -8.6760  3.5915  2.2044 -0.3447
#>   3.6870 -4.2228  0.0170  4.4906 -0.7148
#> 
#> (1,2,.,.) = 
#>   2.9148  7.1475  2.6259  0.7836  2.0463
#>   1.6566  2.2269 -4.2176  0.3844 -5.0686
#>  -1.5057  1.7086 -3.6555  7.3055  0.6776
#>  -0.5608  4.1712 -8.7006  7.5598  4.8863
#>   0.2058  7.4104 -1.6123  1.7842 -4.9439
#> 
#> (1,3,.,.) = 
#>   6.6791  2.6795 -0.6120  1.1870  1.1087
#>  -3.4221 -2.5527  4.1732  1.6172  5.9491
#>  -3.3692 -3.8300  8.0035 -11.5871  2.7627
#>   2.5010 -14.8966 -1.2025 -9.0161  3.9864
#>   6.9829 -2.5833  4.9990  4.2682  1.4561
#> 
#> (1,4,.,.) = 
#>   1.8005 -4.3846 -2.4868  0.3061  2.9196
#>  -1.1223 -7.1243  2.8625 -0.5572  0.7980
#>   2.7798 -3.0581  2.4222  4.4361 -2.8283
#>   0.8278 -1.6823 -8.1614 -2.8767  1.0986
#>  -0.3550 -0.3729 -6.0236 -2.9959  1.8867
#> 
#> (1,5,.,.) = 
#>    2.7292   6.0449  -3.0908  -0.2330   4.0662
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]