Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   -4.9819  15.4201   7.2964  -0.2801  -2.9515
#>    8.5223   0.6689   6.8598   4.2131   1.4642
#>    7.2974   5.8444   2.5709  -3.0454  11.9282
#>    2.4449 -11.5427   0.7447   0.2285   7.3967
#>    8.3833   2.9643  -3.6349  -5.3794  -0.2870
#> 
#> (1,2,.,.) = 
#>   -2.4653  -7.5399   5.8834  -6.0364  -8.9845
#>    4.4188  12.3506  -0.5495   3.7874   2.3367
#>  -10.6945   7.6717  10.8756  -1.8952  -1.6022
#>    3.0461   6.0540  -4.5503  -1.6476   6.4968
#>    1.7559  -4.9923   2.2632  -1.1299  -7.2580
#> 
#> (1,3,.,.) = 
#>  -1.7876  6.3092  6.5608 -0.9300  5.8089
#>  -1.9246  1.0590 -1.0220  1.6432  4.9451
#>   4.7078 -6.7573  4.5426 -7.7238  2.4132
#>  -4.0112  1.5771 -3.7141 -6.4647 -3.5902
#>  -4.3800 -1.1169  2.6540 -1.9023 -0.6960
#> 
#> (1,4,.,.) = 
#>    4.5639   7.7464  -6.4244  -2.2218   1.9135
#>   -3.8612  -3.6067  -3.3017  -2.8934   6.8919
#>   -1.3315 -13.1605   1.7044   5.8895  11.3741
#>   14.0540  -0.7854  12.2961   4.3192   9.6555
#>    0.1648  -1.4318  -3.7480   2.0255   0.9023
#> 
#> (1,5,.,.) = 
#>   2.1588  1.2086 -0.9362 -4.2654  1.1884
#>   5.1673 -1.4304  8.8540  8.5885 -1.2423
#>   4.8948 -7.7570 -0.5604  5.2059  5.0610
#>  -4.7332  7.5978 -6.1148 -9.3669 -0.1263
#>  -4.3766 -9.9737 -7.0490 -2.3554  6.3364
#> 
#> (1,6,.,.) = 
#>    4.2089   1.8841   4.5940   6.0473   1.9857
#>   -9.3275   1.2711  -3.6334   2.2515   1.6834
#>   11.8008  -4.5928 -10.2462   5.3197   5.7052
#>    0.7492  -4.4178   5.6462   4.8558   4.5984
#>   -0.8200  -4.6593   6.5684   9.5202   1.8884
#> 
#> (1,7,.,.) = 
#>   0.4213 -7.4980  0.0837 -2.4890 -3.8955
#>   2.7315  0.3373 -1.7870 -2.2954 -2.5914
#>  -3.5940  4.5722  5.4906  4.2236 -1.0460
#>   1.8401  3.9772 -11.9184 -7.3575  5.0666
#>   4.8394  2.0948 -7.1201 -4.0899  0.1105
#> 
#> (1,8,.,.) = 
#>   -5.2932   7.4397   8.6222  -4.8576   0.7271
#>   -2.0619   4.9121   9.6196  -7.1980  -6.5354
#>   -3.7706  -6.0629   5.0730  -1.4315  -6.4147
#>   12.8932   2.9425   7.1253   9.3403  -1.7983
#>   -2.6609   5.5162   4.7607  -1.7937   5.6835
#> [ CPUFloatType{1,8,5,5} ]