Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.6143   2.8220   4.3484  12.8542  10.9534
#>   -7.0898   5.6533  -1.7187   3.1381  -1.0518
#>   -7.6532  -5.0314  -8.9808  -1.6589  -9.4841
#>    0.8350  -4.6705  11.1024   3.9767   5.6909
#>   -0.8706   7.3704   4.6400  19.7004  -4.7378
#> 
#> (1,2,.,.) = 
#>    1.4142   2.8823  -3.9634   5.3864   5.0051
#>    3.1438   0.3393  -1.1176   0.0548   2.5555
#>    1.3217  -0.2263   5.4554  13.3783  -7.4900
#>    6.6183  -5.0104  -1.1522 -10.1652  -6.3639
#>    0.3199  -8.2034  -0.3581  -7.7796   4.2205
#> 
#> (1,3,.,.) = 
#>    0.1197  -1.2419   3.0110   4.2406   1.5956
#>   -0.1607   9.1214   4.6991  -0.9824  -0.2446
#>    2.8382  -2.7582 -10.8028  -2.1866   2.0420
#>   -3.9233  -9.8055  -1.8206   7.0895  -3.3932
#>    3.8279   0.0724  10.8993  -2.5320   2.6545
#> 
#> (1,4,.,.) = 
#>    4.2383  -3.6772  -6.3039   5.7341   1.3136
#>    2.6193  -1.9012 -12.3492 -10.7850  -0.5075
#>   -2.0862  -6.5252   7.4094  -0.0011   5.7049
#>    3.1709   8.1475  13.9630   1.6448   2.4076
#>    0.4377   6.8046   3.0384   6.0045   1.3388
#> 
#> (1,5,.,.) = 
#>   5.7892 -0.0285  4.0327  4.5085 -0.7988
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]