Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   4.5247 -2.5636  8.5214 -3.9769 -5.9746
#>   2.6690  6.9445  0.0388  7.7454  3.1078
#>   7.7650 -0.9085 -3.6737 -14.7126  9.7160
#>  -11.3353 -3.8239  1.3281  4.7960  6.2711
#>   5.2664  8.3923 -0.1662  3.7873  2.5719
#> 
#> (1,2,.,.) = 
#>    3.2496  -7.4630   0.8823  -0.0813  -4.9909
#>    5.5575 -10.3420   1.3069  -2.4077  -1.1031
#>   -2.9945  11.9156  -6.0696  -3.2610  -6.3540
#>    1.5155   3.2354  -3.1893  -6.2190  11.1127
#>    0.1623   7.1710  -4.9062  -5.1283   0.6922
#> 
#> (1,3,.,.) = 
#>   3.3582 -0.8275  2.2684 -6.7627  3.3128
#>  -1.9869  1.3173 -5.0077  2.7160 -4.2726
#>   0.1176  0.4242 -2.2784  3.2606  5.8648
#>  -9.2244 -1.4194  1.1344 -8.5064 -4.5069
#>   4.5131  0.5613  0.7095  1.5367 -5.3575
#> 
#> (1,4,.,.) = 
#>   6.5396 -0.2769  1.9308 -3.8463  3.5058
#>  -2.4311  6.7615 -1.1723  5.9786 -13.8829
#>  -3.2637 -4.6999 -4.9723 -0.6359  2.7473
#>   0.2097  9.7608  2.0803 -8.3006  1.8584
#>   6.8111 -4.9392  2.8522  4.9001  5.7390
#> 
#> (1,5,.,.) = 
#>    0.0939  -0.0867   3.0771  -0.2569   0.0858
#>   -1.1462   2.4150   1.6543  10.9297   0.0709
#>    3.6044  -3.8082  11.2985 -11.1168   3.9180
#>   -2.3963  -1.2846  -0.8350   4.5580   0.6524
#>    6.6147   2.6445  -5.0096   8.3467   4.3054
#> 
#> (1,6,.,.) = 
#>  -0.3202 -3.4035 -5.1246 -2.0641 -3.6342
#>   0.0828  3.9967  4.7672  0.7800 -1.4425
#>  -0.0390  3.0940 -0.6782 -4.8385 -1.8986
#>  -2.2784 -9.1464  0.4769 -0.8680  2.0762
#>   2.9578  2.4193  4.1523 -1.5880 -10.1609
#> 
#> (1,7,.,.) = 
#>    3.2858   4.9365  -0.7546   8.0840  -6.3284
#>    4.2430  -2.5305   8.3381  -3.1048   4.1082
#>    2.3266   6.8213  -2.5499  -6.3303  -1.9806
#>    7.2764   1.5998  -2.7330   9.2004  13.3941
#>  -11.1460   5.4237  -0.2832  -3.8571   0.4329
#> 
#> (1,8,.,.) = 
#>   -1.1486   4.3550   2.0178   1.2295   5.7356
#>   -3.1822   2.8283  -1.6488   7.7764   1.9796
#>   -1.5174  -8.6087   3.3470   7.1780   4.4484
#>   -1.8043  12.0576   2.7262   6.9968  -3.9764
#>    0.6294   2.8320   0.4024  -0.2191   0.2245
#> [ CPUFloatType{1,8,5,5} ]