Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -3.5541   0.1340  11.5555 -10.1455   5.1898
#>    3.9914   1.6622  -8.4136   8.1750  -9.5323
#>   -1.0458   5.7948  -3.7086  -4.2089  10.6684
#>    6.5773 -13.2337   4.2014  -9.7015   4.7955
#>  -14.4490   1.2789   6.1686  21.5776  -2.6880
#> 
#> (1,2,.,.) = 
#>   -4.3907  10.2363 -11.8444   3.3066  -3.0480
#>    3.5979  -0.6333  -5.7375  -6.8155   5.6770
#>    1.0358  -1.7718   8.3280   3.1905   6.2463
#>   -3.5157   4.5980  -6.6943   7.6189  -5.7677
#>   -6.5348  -3.5216   4.6577  -1.9670   4.5680
#> 
#> (1,3,.,.) = 
#>   0.7852 -0.1301 -7.0692 -0.2198 -3.3151
#>  -5.0652  6.4705 -3.4053  3.1551  6.4885
#>   1.4688  4.8634  6.9644 -7.9324  4.6831
#>  -2.5768 -2.4689 -8.8636  0.4202 -4.3524
#>  -2.7660  0.7899 -1.9714 -5.4197  3.1038
#> 
#> (1,4,.,.) = 
#>  -5.0990  9.7017 -1.5631 -7.6446 -0.0443
#>  -4.9228 -1.7760  5.8776 -10.5836  7.3651
#>  -1.5679  0.0830 -1.4304 -1.7835 -1.2275
#>  -3.7858 -0.0412 -12.5883  5.6978  3.5045
#>  -0.2855 -4.4517  8.3601 -8.5574  1.6679
#> 
#> (1,5,.,.) = 
#>  -0.6066  6.8681 -6.4750 -3.1575 -5.2515
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]