Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   1.6084 -3.4760  4.9442 -1.3845  1.8975
#>   1.7293  6.6989 -2.8882  0.4966  1.4817
#>  -0.2804 -10.2550  7.2247  0.1894 -4.1485
#>  -0.9585  2.0498 -4.2209 -9.7722 -0.8183
#>   2.9561 -3.9228 -0.6107  0.1121 -0.8458
#> 
#> (1,2,.,.) = 
#>   4.7759 -1.9774 -6.4085  1.8098  2.5684
#>  -5.3117 -4.2276  4.0629 -2.6649  4.4999
#>   8.8466  3.0768  7.8340 -2.4529  0.4341
#>   4.7648 -4.1504 -7.8764 -4.6141  6.0449
#>   0.3317  1.2971 -4.2739  5.3548 -1.5957
#> 
#> (1,3,.,.) = 
#>    4.7941   1.7374   1.6986  -0.6278  -0.1873
#>   -0.9803  -3.5496  11.6338   6.4688   2.1949
#>    5.7949  -3.5175   0.2033   1.9678  -1.6372
#>    1.6885  -6.4527  -1.2045  -1.4585   0.6239
#>   -4.1695  -4.8260   5.9417  -4.5571   4.8039
#> 
#> (1,4,.,.) = 
#>   -1.9137   5.3849  -1.2173 -11.9908  -2.7591
#>    0.7621   3.1297   4.1437   1.4596  -8.2257
#>   -1.6562   2.0199  -4.6867  -9.1243  -2.2696
#>   -0.9100  10.3501  -8.8508  -5.8140  -0.5370
#>    1.1513   4.4225   4.3226  -3.0215  -6.4409
#> 
#> (1,5,.,.) = 
#>   -0.1040  -0.1033   2.1559   0.4238   5.8815
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]