Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -3.0107  2.6936  8.1035 -8.6643  4.3891
#>   2.2687 -14.5017 -6.3096  3.7539 -0.5046
#>  -5.0892  1.6683  3.5089  6.3103 -1.7491
#>   0.3358  0.5173 -1.7769  2.6625  6.4258
#>  -1.5031  4.0863  6.5195 -4.3026 -2.9908
#> 
#> (1,2,.,.) = 
#>    0.6978   5.6053   1.9356   0.2673   7.5423
#>   -2.8250   5.1866   4.5877  10.8132 -13.6531
#>   -3.2114  -2.0764  -9.9918  -1.5124   2.8339
#>    0.8964   4.8960   9.0674   0.3958  -2.3072
#>   -4.0554  -6.5812   0.0381   5.7629   5.8394
#> 
#> (1,3,.,.) = 
#>   2.1740  3.8659  1.6881  1.4812  5.0924
#>  -4.4150 -3.2243  3.7819 -0.5216  3.4326
#>  -1.0521 -1.1335  3.6444 -5.2629  0.3667
#>  -5.4841 -1.1688 -5.0616  9.6486  7.0773
#>  -0.5157  2.6908  5.5346  0.1290 -1.4847
#> 
#> (1,4,.,.) = 
#>   -0.0763   5.6380   5.6527 -10.6461   5.5123
#>    3.1761  -0.2752   3.1726  10.4850   2.5933
#>    4.6298  -0.9594  -3.7696  -5.4753   0.2907
#>    4.3042  -2.1716   2.4378   3.7930   1.8757
#>    8.5376  -2.4329   0.0452  -5.7029   1.1068
#> 
#> (1,5,.,.) = 
#>   3.2163  2.3177 -2.0393  9.7789 -1.1514
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]