Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    2.2176  -5.8345  -1.8880   6.3114  -1.5370
#>    2.1718   4.2041   2.3940  -0.2030  -4.8366
#>    0.9334   6.3011   1.2620   1.8764   5.2892
#>   12.3241   4.6636  -0.2072  -0.3904   0.5950
#>    5.3774  -1.9787   0.6011   4.2020   0.0159
#> 
#> (1,2,.,.) = 
#>   2.1784  6.5130 -6.2456  5.8970 -2.1626
#>   0.5804  7.0183  4.7296 -1.3214 -0.2059
#>   4.7738  2.4388 -7.2965  4.0155  1.5872
#>   6.6335  7.3526  4.5877 -1.9349  1.1081
#>   3.3799 -1.1880 -7.8268 -7.7664  0.2255
#> 
#> (1,3,.,.) = 
#>   6.3211 -6.4212  1.3741  0.9843 -0.9183
#>  -5.5355  4.7472  6.2253 -9.5499  1.9842
#>   4.0450 -2.1036  2.6976 -0.6012 -0.9722
#>   0.1781  6.3567  5.8495  1.1594 -8.1266
#>   1.0849  5.2977  1.6611  7.0236  3.2572
#> 
#> (1,4,.,.) = 
#>    0.8957  -3.0792   5.4038  -1.4378   5.9207
#>   -0.8185 -10.8808  -1.4467 -11.5982   3.7686
#>    0.6290   8.9978  -0.6108   9.0198  -3.3319
#>   -0.0660  -7.5567   7.7348  -6.1848  -2.2490
#>   -0.6869  -6.2114  16.6578  11.0408  -4.5077
#> 
#> (1,5,.,.) = 
#>    1.3847  -0.8084   1.2576  -1.0469   3.4807
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]