Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    1.5450  -0.1299  -2.7065   5.1321   3.3125
#>   -4.1409  -8.3410   1.9553  -3.3304  -2.2018
#>    0.4138   2.1155   0.8999  -0.0725   2.3969
#>   -2.9104  -0.6886  13.1445  -9.0310   2.2215
#>    2.7069   0.3577   0.8432  -2.8648  -0.4802
#> 
#> (1,2,.,.) = 
#>   1.0892  0.5937 -3.3013 -2.5759 -0.1734
#>  -0.5845 -3.7452  1.6358 -0.0439 -2.7466
#>   6.8814  2.6869  7.6323 -10.0282  2.0593
#>   1.3586 -8.5905  1.5439 -10.8203 -1.7988
#>   1.1455 -2.4382 -5.7741 -8.0630  0.0249
#> 
#> (1,3,.,.) = 
#>   5.2437  8.3911 -4.6034  1.6025  0.0096
#>  -0.1740  6.1127 -2.5934  2.0414 -7.7545
#>  -4.4266  8.0505 -5.0359  2.6022 -0.9019
#>  -11.9747 -3.5390  0.5294  1.4066 -0.0329
#>  -2.7113 -0.2886 -10.5046 -4.8810 -1.5061
#> 
#> (1,4,.,.) = 
#>    5.7583  -8.9794  -4.8830   2.8914   1.7922
#>    2.3427 -11.9150   0.5788 -10.6110  -3.0845
#>    6.3824  12.9411  -9.5175  -1.8389   4.3841
#>   -9.0548  -1.0787   0.7505   1.2792   7.1082
#>   -1.6317 -11.3894  -8.9768  -7.4410  -6.7142
#> 
#> (1,5,.,.) = 
#>   -2.0158  -2.0200  -0.4335  -3.6228   2.4701
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]