Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -1.1012  5.5839  1.8122  0.8258 -7.3393
#>  -2.1950 -1.8602  2.8601  6.6617 -2.3251
#>   4.1597  2.5679 -3.9130 -0.2439 -4.2846
#>  -2.1240 -5.1318  2.2491 -6.6270  7.9282
#>   4.8318  1.1933  7.0275 -1.2019 -1.6025
#> 
#> (1,2,.,.) = 
#>    5.8042   0.9000  -2.5795  10.8978  -4.9126
#>    6.9631   3.0681   6.0450   1.6698   8.3346
#>    5.9873   5.5173  -0.0483 -10.7899  -1.0211
#>    2.9545 -12.5700  -2.5009  -8.8733   5.1434
#>    1.9054  -4.2049  -2.6949   0.5966  -8.2003
#> 
#> (1,3,.,.) = 
#>  -0.8560  1.3400 -6.0400 -3.6040  5.5857
#>   3.4883  1.2494 -7.9900 -1.0955 -1.5007
#>  -0.7039 -8.2752 -3.1481  6.9932 -7.4125
#>   1.8730  2.8466 -0.3726  2.1923 -2.2347
#>  -3.9589  1.5581 -2.5740 -6.2123 -0.2614
#> 
#> (1,4,.,.) = 
#>   -1.8405   0.3266  -4.9824  -1.7248   2.8302
#>   -1.5687  -2.1843   0.4188  -5.5145  -2.0998
#>    9.0726   3.0494  -0.4360  10.6919  -1.4521
#>    2.9984   4.1762 -11.7169   1.3491  -4.0201
#>   -3.4309  -0.2874   0.3173   1.1203   2.6208
#> 
#> (1,5,.,.) = 
#>   0.2031  1.4480  3.1446  5.2531 -5.9000
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]