Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    3.9421  -2.7716   6.4478   3.0279  -7.8044
#>    1.0534   8.4491  -4.1091   3.0606   6.2162
#>    0.2559 -14.7836   2.6502  22.6987  -2.1501
#>    3.2447  -4.5661   7.9644  -2.2879  -5.2302
#>   -6.3045  15.2665  -5.2219   1.4747   4.8622
#> 
#> (1,2,.,.) = 
#>   -4.2954  -2.8950   2.7619  -4.9995 -11.6727
#>   10.0772   6.5658   6.2568  -1.2067  -2.0788
#>    8.0195   7.8704  13.4699  -4.1526   2.8158
#>   -2.7300  -5.5696  -3.7574  -0.8195   9.1970
#>    2.7586   2.2808  -3.8216  -1.1150   4.9985
#> 
#> (1,3,.,.) = 
#>    5.2361   7.2294  -4.9662  13.0389  -4.4024
#>   -1.6612  -5.9017   2.1110  13.8624  -3.5783
#>   -0.8705  -6.3926  -0.0729   3.0561  -6.6180
#>    5.0075  -4.9633   3.2353  -8.5220  -3.1180
#>   -4.9541  -1.9341   7.2629   4.2547  -3.1501
#> 
#> (1,4,.,.) = 
#>    3.0037   6.0015   1.1353   0.7004   1.1304
#>   -1.5847  -4.9246  -7.0868   2.6243   2.7751
#>   -7.9582  -4.9536   8.1043 -15.0214  -0.3329
#>    5.4938   9.9232  -9.6448   1.2053  12.1352
#>    4.3038  -3.5006   7.0383   4.6821   6.0879
#> 
#> (1,5,.,.) = 
#>   -4.3459   3.4306   8.7031  -0.3631   8.2612
#>    1.4519  -5.1385  -9.0213   1.8501   4.9133
#>   -2.8035  -7.6205  -3.0344   5.5390  -1.7051
#>  -11.6210   3.2170  11.1163  -3.9823  -7.7165
#>    5.7409  -9.9262  -5.1490   1.1958  -2.2574
#> 
#> (1,6,.,.) = 
#>  -1.6674  0.5367  8.6139 -7.6563 -3.2011
#>  -0.2719  4.1356  7.2939  1.8638  4.7427
#>   2.8846  0.7575  4.6594 -4.6902 -1.5745
#>  -10.1149  2.9447 -3.9891 -8.9825 -0.6310
#>   7.7687  0.3345 -6.8998  7.0819  1.5945
#> 
#> (1,7,.,.) = 
#>    2.5410   0.0307   3.9461  -2.2793  -5.2953
#>  -10.6706   1.1420  12.9382   2.5283  -2.7881
#>   -3.5774  -5.7784   1.9396   0.0802  -1.9985
#>    6.7545  -1.6455 -12.1507  -3.2365   7.3346
#>   -2.0992   5.5099   5.6053   3.2151  -6.9274
#> 
#> (1,8,.,.) = 
#>    1.1419  -3.7804  -2.3402  -8.9931   9.2812
#>   -0.1031   6.6184  -1.5618   2.8972  -0.8986
#>    0.5048   2.3964   0.2656  15.0013  -7.1948
#>   -5.0651 -15.3735   2.2232  18.2544  -1.8101
#>    2.9777  -6.8105   1.1220   3.0700   5.7919
#> [ CPUFloatType{1,8,5,5} ]