Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    6.2290  -3.5145   2.6597   2.8240 -10.4595
#>   -3.7757  15.0257  -4.4488  -4.5800   6.1128
#>   -4.4282  -6.1300  -3.4805  -9.5114   0.9564
#>    1.7146  -4.3512   2.7530  -7.1486   3.8231
#>    0.2193  -3.4697  -0.1021  -2.0025   3.8457
#> 
#> (1,2,.,.) = 
#>  -0.7628  6.1105 -11.1653  8.6405 -3.2801
#>   3.1322 -6.1904  6.6671 -4.3939 -11.8475
#>   2.9596  2.2877  3.2862  1.8960  4.6523
#>  -10.8447 -3.4056 -4.4288  3.5061  4.0787
#>  -3.2590  1.6459 -2.9759  1.2824  1.8097
#> 
#> (1,3,.,.) = 
#>   4.0448 -9.8096 -1.0129 -5.8272 -3.2374
#>   6.4920  5.1316  1.2761  3.5249  1.6707
#>  -0.8867 -7.4459 -9.9853  0.8563  1.9868
#>  -0.0667  2.0311 -3.0603  1.5513  5.9617
#>  -2.1265 -0.8882 -1.7908 -1.7067  5.5568
#> 
#> (1,4,.,.) = 
#>   10.8507  -9.7726  -3.3531   0.3298  -1.1506
#>    4.7653  -1.5859  -1.9893   7.1925   2.0269
#>   -9.6811  -1.9828 -11.7392  -2.9589   3.5099
#>   -4.5478   2.8205  -0.5366  -0.0671   6.9125
#>   -1.1469   2.7287   6.5938   3.8296   2.3512
#> 
#> (1,5,.,.) = 
#>    1.3521  -7.6518  -0.4713  13.5282   6.7189
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]