Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -4.0870  -3.8187   5.1422   2.5038  -2.2057
#>   17.1827  -2.5129  12.5472  -7.3937   3.7540
#>    0.9865   7.2586  -1.7575   6.4757   6.9222
#>   -4.7537   7.5623  -2.4539 -18.2069   2.2940
#>    3.0532   9.0403  10.3569  -2.3142   0.9239
#> 
#> (1,2,.,.) = 
#>    0.9182   1.6685   5.7290   0.0917  -3.0335
#>    5.1323   0.7410   2.5337  -6.2594   5.5359
#>   -2.4353   2.6695  -2.5308  10.2397   0.8164
#>   -1.3408  -3.8704   3.8643  -5.9544  -5.8999
#>    5.6337  -0.7972  -2.5412   3.9248   1.2220
#> 
#> (1,3,.,.) = 
#>   9.1799 -9.7226  0.4381  2.3396 -2.1883
#>  -7.4040 -7.5536 -7.0918 -9.0761 -6.0713
#>   2.3117  1.9960 -1.4266 -3.4201  1.5055
#>  -2.4380  9.6356  7.1876  0.3920 -4.6870
#>   1.7998  0.1010  0.3287 -8.3443 -2.9262
#> 
#> (1,4,.,.) = 
#>    0.6562  -2.8209  12.1939  -8.3545   4.7894
#>    0.3731  -6.0473  -2.8638   3.3098   8.2337
#>   -4.0191  -2.5243   6.6531  -1.3686  -0.9472
#>   -2.1100  -2.8480  11.3525  -3.7215   0.5796
#>   -1.2614   2.4180   0.5294  -8.4628  -2.6867
#> 
#> (1,5,.,.) = 
#>  -9.1938 -4.5195  0.3214 -6.3528  1.2671
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]