Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   0.1432 -0.5213 -3.0202  0.2099 -5.3991
#>  -0.2656 -8.2436 -3.6631  1.9306  0.2609
#>  -0.8609  1.4368  5.2135  5.5014  4.2359
#>  -7.4000 -9.0311  9.8476 -0.1434 -4.3547
#>  -7.3651 -4.3697  1.2032 -7.4791 -6.8706
#> 
#> (1,2,.,.) = 
#>    2.7864  -3.8582   2.1393   3.0295  -1.3342
#>    6.9223   2.5283  -8.6640   2.8985  -3.8147
#>   -2.9596   5.9107   1.9300 -10.7781  -2.3143
#>    3.4366   9.9201  -2.5027   0.3437   9.3606
#>   14.2538  -2.9165   7.5233   1.5669  -4.6529
#> 
#> (1,3,.,.) = 
#>  -1.5918 -0.1518 -2.7988  0.7587  4.5224
#>   2.0845  1.7842 -1.0887  3.0612 -2.0474
#>  -2.4189  1.1326  0.4418 -7.3844 -2.1778
#>  -2.0134 -3.6621 -4.1564 -1.7099  0.7997
#>  -1.5657 -4.7098 -0.7890 -0.9499  1.4993
#> 
#> (1,4,.,.) = 
#>  -5.5694  2.0511  3.7306  4.3671  2.5741
#>   5.9582 -13.7228  3.7673  5.6279 -4.1325
#>  -4.4002 -3.7703  0.4996  3.4551 -7.0937
#>   0.3782  3.8863 -3.7536  2.2595  8.4656
#>  -2.3951 -0.3750  7.3519 -2.7878  1.5883
#> 
#> (1,5,.,.) = 
#>   2.3707  5.8181 -0.6393  0.4223  4.6258
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]