Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    0.6470  -6.4223  -2.8970  10.7745   4.0952
#>    2.3167  -0.8395   9.6859   9.4134  -3.0314
#>   -1.3487  -7.6222  -5.1335  -7.1064   2.2236
#>   -1.7429  -6.5802   7.4414   5.8215  -3.5750
#>    1.2169  -2.0421  -1.7012  -1.7909   0.8993
#> 
#> (1,2,.,.) = 
#>    4.7590 -11.9310   2.5782   8.3073   4.2833
#>    2.9997  -0.3005   2.2356   5.0668   5.8578
#>   -5.0408  -2.1689 -12.6740  -5.6290  10.2270
#>   -7.4248   2.7269   0.9931   1.8122   0.6790
#>    0.7040   1.7865  -1.5373  -9.3323   6.4639
#> 
#> (1,3,.,.) = 
#>  -3.4069 -2.3124 -3.2604  6.7098  2.7603
#>  -3.7613  4.9487 -9.7047  2.7213 -6.4795
#>  -3.1620 -10.1235  0.8183 -2.9992 -2.4964
#>  -1.3463 -5.1666  1.5078 -2.6464 -1.5033
#>   2.3637  4.5988  1.0593  2.6394  2.3030
#> 
#> (1,4,.,.) = 
#>   -4.6917  -3.1956  -3.5073   3.5459   1.0111
#>    0.1401   3.6252  -9.1027  10.9072  -0.3876
#>   -5.7708 -13.3670  -1.5686  -4.4695  -7.0339
#>    5.9042  -8.1382   7.6566  16.0927  -0.6766
#>    0.1269   1.2716  -4.7544  -0.6452  -1.6976
#> 
#> (1,5,.,.) = 
#>  -0.0454  0.0611 -1.9819  8.4745 -4.8406
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]