Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    0.0300   7.4449   4.6892  -2.2334   4.2230
#>   -4.0881   0.1245   0.1433   0.2543   0.0834
#>   -2.5053  -6.6914   2.1014  -6.0802  -1.2483
#>   13.4086  -3.0649  16.0991 -11.9728   8.7518
#>   -0.5192   2.8714   2.7728  -3.3530  -0.5693
#> 
#> (1,2,.,.) = 
#>  -5.0467 -0.0045 -1.9210  1.5617  2.6934
#>  -5.0193  3.1334  4.8816  2.7719 -0.7533
#>   1.9840 -1.4401  0.5064  1.4156  0.1394
#>   3.9252 -8.3830 -9.1683  0.9204 -6.9798
#>   1.4028  0.5552 -7.2441 -4.6886 -6.7583
#> 
#> (1,3,.,.) = 
#>    2.8854   1.1765  -0.9101  -4.1425   1.2495
#>    3.9187  -2.9141  -4.6189   4.4326   7.7865
#>    0.5179   0.3222   3.3634   2.5794   0.7010
#>   -7.3339   0.6343  -2.6699  -8.7706  -1.2127
#>    3.0565   2.6093  13.4756  -4.4089  10.4744
#> 
#> (1,4,.,.) = 
#>  -6.0218 -8.1147 -7.9057 -3.8192 -1.7183
#>   5.0859  3.6648 -6.8533  1.7164 -2.1265
#>  -1.4993  0.8420  1.8748  3.4103  8.2170
#>  -0.2418  5.5899  6.8834  3.1730 -0.3869
#>   6.2668 -10.7327  7.0911 -2.4030  1.9445
#> 
#> (1,5,.,.) = 
#>  -5.9667  0.0679 -0.7351 -2.1603  4.2531
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]