Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    1.5297   2.4223   0.8714  -3.8710   5.6666
#>    5.9453  -1.5504  -2.5708  10.1308   5.6582
#>   -0.0242  -6.2957   2.6019  12.3505  -2.2625
#>   -4.1604   3.7008   3.0038  -1.7760  -5.4959
#>    0.6925   2.6220  -5.6152   1.3031   3.2958
#> 
#> (1,2,.,.) = 
#>    5.6636  -3.3054  -2.7405   6.7862   4.6309
#>    3.1597  -6.1737 -15.9276   3.7868   6.3714
#>   -6.6701 -11.7130   5.4714   7.8319  -2.0994
#>    0.6403   2.7219  13.9254  -1.5249  -4.7972
#>   -1.0335  -0.2952  -2.6980  -6.9202   4.4785
#> 
#> (1,3,.,.) = 
#>   -4.2426  -1.3773   6.7982  -3.9848 -11.5494
#>   -6.5685   4.1002  13.5306   6.8218   0.4420
#>    9.0605   2.8615  -0.8279  -3.0143   4.1519
#>    1.5769  -2.5728   3.3736   5.5297   4.1609
#>   -5.0542  -3.6299   5.4143 -11.7264  -4.5783
#> 
#> (1,4,.,.) = 
#>    3.3355  -0.6999   3.2744   0.4107   0.0881
#>   -3.8698   3.1812  -1.0439  -9.2540  -1.9492
#>   10.9136   7.5500  -3.1030   0.2495  14.1525
#>   -0.4616   2.7505  -1.0486   8.9183  -3.9921
#>   -2.1117  -4.1024   2.1391  -0.3705  -1.7502
#> 
#> (1,5,.,.) = 
#>  -0.2318 -4.7205  5.9795 -2.0744 -2.4217
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]