Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    1.9712  -0.6488   3.4260   2.4139  -2.0340
#>    8.5507   9.7628  15.7296   2.6138   2.0149
#>   -4.9540  -0.7125  -8.5744   3.3866   4.6238
#>   -3.2880  -2.0031   0.9344  -1.0248   1.3352
#>    4.0846   6.4950  -0.8770   2.8620   3.7037
#> 
#> (1,2,.,.) = 
#>  -2.9499 -6.2838  2.3626  4.7274 -4.2754
#>   1.0402  0.5408  2.2402 -1.3538 -0.0236
#>   4.7592 -12.5851  7.8036  2.3539 -4.9718
#>  -6.2660  4.7854 -13.1924  1.5023  2.1499
#>  -10.5197 -7.1269  5.3832  3.2662 -4.8747
#> 
#> (1,3,.,.) = 
#>   6.3176 -1.5136  1.7543 -0.6215 -4.1217
#>   1.5915 -5.9707 -0.5998 -9.3549  1.5442
#>  -0.2512  3.8220 -12.4825  5.6162 -0.6523
#>  -3.2690  4.5777  3.3527  3.7999  4.4292
#>   0.7060 -6.6069 -2.1562  0.7483  5.3432
#> 
#> (1,4,.,.) = 
#>   -1.1574   0.0489  -6.1699  -3.2137  -1.2370
#>    0.5621 -17.2945   7.9332 -11.4452   0.3776
#>    9.0467  -9.6862  11.6008  15.4611  -0.0525
#>   -6.8355   4.9849  -3.7762   6.6653  -8.8637
#>   -0.4807   8.7882  -9.5010   5.7015  -0.6556
#> 
#> (1,5,.,.) = 
#>   -3.0571   2.3561  -8.1299   0.6114  -2.7725
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]