Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   -6.7352   4.5207   3.1809  -3.2571  -8.8025
#>   -0.6917   4.3741   9.4463  12.4223   4.8701
#>    1.8166  -0.2720  -5.5085 -12.7067   2.1776
#>   -8.6747  -2.8475   5.7520   4.3040  -3.0128
#>    5.3373   1.9571  -1.1926  -7.2064   2.6974
#> 
#> (1,2,.,.) = 
#>   3.6328  7.1966  6.2769 -4.1961 -8.9584
#>   1.0518 -5.4110  6.1148  4.6490  4.0348
#>   4.2070  4.6418 -4.5894  2.4298  0.7264
#>  -0.1213 -7.7344 -11.1384 -3.7803 -5.2400
#>   3.6290 -2.6493  0.4713  5.6207  5.0020
#> 
#> (1,3,.,.) = 
#>   6.9942 -5.6425 -8.9214 -3.5571  2.2214
#>  -4.0623 -2.8171  7.5024  7.5474  0.8746
#>   2.4144  0.6087 -5.4951  1.6702  3.7909
#>  -0.9200 -1.7780 -2.7462  3.6730  2.2886
#>  -1.1750  3.8700  3.7342  0.3783  1.2824
#> 
#> (1,4,.,.) = 
#>   3.8779  0.1967 -4.4352  2.6424 -0.8010
#>  -3.6309 -7.0595 -6.9523  0.0672 -2.9846
#>   3.0640  8.2980  4.2321  5.5741 -0.8411
#>   5.2825 -6.5650 -6.7520 -6.0906 -1.9723
#>  -3.1665  1.7242  0.8445  4.4522  1.2617
#> 
#> (1,5,.,.) = 
#>   1.3917 -1.4185 -3.6496 -6.0074  0.6024
#>   3.3281  3.8187 -2.5331 -3.0597  6.3589
#>  -4.8020  3.3434  6.3879 -4.2468 -5.5208
#>   1.3408 -1.3616 -13.2736  4.7146  7.4853
#>  -1.5069  0.7438  9.5950  0.7343 -0.5171
#> 
#> (1,6,.,.) = 
#>   6.0374 -0.9968  3.2415  2.4045  1.0893
#>   1.8273 -1.9829 -0.9077  0.9950 -3.7424
#>  -5.1596  5.8034 -5.6752  1.4541 -6.1392
#>   7.8612 -3.8173  1.3246  5.9785 -0.6011
#>  -2.4237 -3.2377  3.5633 -0.8885 -3.3294
#> 
#> (1,7,.,.) = 
#>   1.0343  2.8307  0.6409  0.9831  1.9679
#>   2.5620  5.3542 -0.6069  0.2226  3.7899
#>  -2.9894  0.7568 -5.5890 -2.8741  0.4750
#>   0.8378 -4.2920 -5.4254  6.1291  4.2945
#>  -1.2909  6.1218 -0.3022 -0.8540  1.2561
#> 
#> (1,8,.,.) = 
#>   1.1358 -7.5325 -0.8255  1.8853  1.8479
#>   1.5711 -1.2647  8.3031  5.0166  6.3347
#>   1.8183 -7.1348 -0.5809  2.1161 -3.2856
#>   1.8227 -13.0477  1.2087  8.6065 -3.0948
#>   5.7092  0.9823  4.6161  6.3981 -6.0932
#> [ CPUFloatType{1,8,5,5} ]