Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -0.0183  1.5949  1.4855  0.1630 -1.2824
#>   2.5913  3.2195  4.6402  0.4546 -4.3613
#>  -1.7385 -9.2862 -1.4986 -8.4850 -1.4684
#>   0.5933  6.0077 -0.3904  4.9700 -9.8046
#>   0.7781  4.3149  6.9415  1.4005  5.2556
#> 
#> (1,2,.,.) = 
#>   -2.7029   0.2626   5.5444  -0.1342  -0.3909
#>    2.5886  -1.7963  -0.5599   5.6572  -1.8149
#>    4.1622  -7.3990  -3.9713  -7.8727  -1.3965
#>    6.3126   5.9926   2.4177  -3.0840  -2.3620
#>    5.8449  12.4896  -6.3167  -2.9365   8.9168
#> 
#> (1,3,.,.) = 
#>  -0.7435 -2.9778  6.8988 -2.8970 -3.9008
#>  -3.0848  4.7433  1.8181  2.5851  0.8891
#>   1.1160  0.0800  0.8332 -4.6290  2.2100
#>   6.3431 -1.8937 -9.1749  4.9253  0.1841
#>  -0.6676 -6.2148 -8.0356 -2.0113 -1.0772
#> 
#> (1,4,.,.) = 
#>    1.4354  -3.6708   1.7269  -1.1703   1.3516
#>   -0.3615 -11.0344  -4.2016   4.5205   1.4705
#>   -5.8952   2.7427  -4.8169  -3.8018  -1.3066
#>    6.1092  11.3700   9.1047  -4.7770   2.6070
#>   -1.6547   7.5862  -1.3354  -4.6152  -1.8780
#> 
#> (1,5,.,.) = 
#>   1.4812  2.1389  1.2210  3.0367 -3.0243
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]