Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.2861   3.5740   5.3716  10.0311   2.9481
#>    3.1102   1.4960   3.4912   2.0238  -1.8013
#>    4.1440  -5.1116  -1.1478   5.3780  -0.5305
#>    7.8812   0.7207  -6.5483  -2.0946  -4.1962
#>    0.6270   4.7506 -17.9693  -7.8853   2.9849
#> 
#> (1,2,.,.) = 
#>   0.1477  1.5936 -4.1467 -0.0023  1.7161
#>  -0.0403 -1.4357 -3.1531 -6.9927 -0.5278
#>   1.0244 -2.5721  5.9671  4.3883  2.6848
#>   2.2131 -0.7075 -2.6415  4.2446  0.2362
#>  -3.3993  3.2425  2.9011 -3.5323 -2.5270
#> 
#> (1,3,.,.) = 
#>    1.6421  -6.7055  -0.5279  -0.6853   4.0748
#>   -0.8381  -4.7646  13.8211  -4.4893   2.3427
#>    3.2290  -2.4457  -2.5484   4.4655  -2.1516
#>    6.6550  -6.1253   0.3803  24.9023   6.9434
#>   -2.1883   4.0248   5.3342  -6.1260  -5.5562
#> 
#> (1,4,.,.) = 
#>  -3.3504 -0.5610  0.4137 -2.0976 -2.5791
#>   6.8906 -13.3459 -7.3050 -5.7603 -7.7392
#>  -0.4924 -1.0725 -1.9572  7.4428 -5.8285
#>  -0.8989  6.2468 -2.0086 -2.8552  5.5528
#>   0.7372 -1.6019 -0.3256  2.4296  0.7556
#> 
#> (1,5,.,.) = 
#>   9.9078 -9.9184  3.0087  1.6862 -7.6349
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]