Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   1.9068  6.1698 -4.9156  1.0613 -7.6183
#>   4.4856 -2.3910  0.6572  7.3670  2.4091
#>   4.6711 -1.7061  0.3564 -0.8683 -2.1314
#>  -3.1261 -2.7618  1.8228  5.4772 -0.3422
#>  -4.4392  0.9073  0.4555 -8.3045  3.7115
#> 
#> (1,2,.,.) = 
#>  -5.3136  7.3471  3.0439 -1.6026  1.1202
#>  -7.0252 -4.6551  0.1893 -8.7512 -0.2692
#>   0.7481 -2.8000  2.3464 -6.3770  5.3306
#>   0.4484  4.4186  0.1867 -0.8572  3.1332
#>   0.7827 -2.2620  3.7353 -5.1762  2.0003
#> 
#> (1,3,.,.) = 
#>   2.0026  3.4145  6.4648  3.5997 -9.7786
#>  -2.7493 -5.5796 -15.8857 -7.2643 -6.6882
#>  -3.0808  0.3806  2.1299  0.8432  3.4009
#>   2.5315  0.4560  9.5449  7.4246  1.2211
#>   1.3222 -9.5882  5.3305 -0.6182 -5.3189
#> 
#> (1,4,.,.) = 
#>  -4.8943  8.6793  2.4820  9.7242  0.2645
#>   3.2832  0.8070 -1.7733 -16.5891 -10.8029
#>  -3.4499 -2.4347  4.5125 -3.9430  6.5934
#>  -2.4046  2.5657 -1.0782  5.8932  0.0306
#>  -3.8376  0.2102  6.3786 -4.6217  7.6513
#> 
#> (1,5,.,.) = 
#>    5.4871   0.5056   4.9282  -2.6127   6.7587
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]