Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -1.6475   6.8004  -9.1702   1.5122   4.4050
#>    0.5660   4.4753   3.3208  -4.0748 -10.3297
#>   -6.3141   0.9981  11.8099   6.8970   0.5841
#>   -2.4069   0.7975   2.9161   2.3913   1.0941
#>    2.6240  -3.7027  -5.4541  -5.4291  -0.2425
#> 
#> (1,2,.,.) = 
#>   3.5349  7.5646 -4.5150  1.5448 -3.7963
#>  -0.6360 -5.2424  6.0377  2.5786  1.1375
#>  -4.3970  5.4978 -2.9108 -4.6892  4.1260
#>   1.4572 -1.5708 -7.3152 -1.3745  0.2438
#>   0.3015 -1.9462  0.1035 -6.7784 -0.4954
#> 
#> (1,3,.,.) = 
#>   0.5530 -6.1854 -2.6959  8.8010 -0.6814
#>  -3.5878 -1.5224  1.8602 -10.9415 -1.9309
#>   7.1322  2.6990 -8.4936  4.6826  0.2362
#>  -5.9269  2.5314  8.2141  0.4381 -2.2072
#>   1.6425 -2.0662  2.1614 -1.8252  0.4704
#> 
#> (1,4,.,.) = 
#>   0.0156 -7.1524  6.7071 -4.9212  1.9484
#>   1.9323 -4.8641  4.3529  3.7664  0.2345
#>   3.9015 -7.5332 -2.3891  2.1768 -1.2262
#>   1.5634  0.4797 -0.2641  1.8262 -1.4684
#>   6.8533  5.5249  3.5591 -4.1976  4.6379
#> 
#> (1,5,.,.) = 
#>  -0.7293  2.1886 -3.2256 -3.6512  1.3584
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]