Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   3.8377 -4.5959 -5.9536  7.8659  4.6091
#>  -2.2737  3.8183  5.3598  7.1267  4.0081
#>   3.3859  0.8566 -0.5117 -1.0650 -1.0787
#>  -3.2845  8.3047 -1.3939 -7.7463 -2.6836
#>  -4.6514  2.1295 -1.8052 -4.4530 -2.5349
#> 
#> (1,2,.,.) = 
#>   -0.8791   9.6916   8.2581  10.3969  -5.1041
#>   -4.6805   0.5242  -2.0084   1.8730   1.3811
#>    1.3117   6.5270   3.3286  -5.5941  -4.6068
#>   -5.9978   1.2744   7.3246   5.6317   0.9536
#>   -4.1553  -0.9994   8.2206   7.7047  -1.3331
#> 
#> (1,3,.,.) = 
#>   -0.8872  -6.4665   2.8829  -4.6767  -3.5756
#>   -0.1081  -0.6451  -8.2975  -7.0287   0.9092
#>   -3.1887   0.3526  10.4040   8.7053   6.7948
#>   -0.8875  -9.1147  -8.4719  -4.6452   1.2826
#>    1.6134   8.3716   3.5398  -8.9942   8.6217
#> 
#> (1,4,.,.) = 
#>    5.3300  -2.4602   0.7514  10.7365  -1.3553
#>   -3.4203  -2.9066  -7.4540   6.1639  -0.4074
#>    1.5817   2.0030  -2.6580 -11.8014  -4.5332
#>    0.3051   3.3836  -3.4086 -11.7864  -4.4811
#>    3.8035   0.4293  -6.1705 -10.7286  -1.2149
#> 
#> (1,5,.,.) = 
#>   3.1217 -4.3784 -0.8317 -4.0023  0.9723
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]