Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   1.7440 -5.1929  0.3235  0.3050  1.1386
#>   4.8506 -9.0011  0.7451 -3.7966  2.3083
#>   8.1775 -3.8996  6.9590 -7.2071 -6.5069
#>  -6.0185 -6.0147  1.6683 -2.6515 -6.3836
#>   2.3426 -0.4941  4.1287  3.0150 -1.3402
#> 
#> (1,2,.,.) = 
#>   1.5553  4.2408 -4.4912  1.6076  2.8381
#>  -4.6925  7.5617 -2.7786  5.8044 -0.6175
#>  -4.9512  9.3908  1.8972 -3.3354  6.0474
#>  -3.7048  2.4108  0.4961 -3.4415  3.2842
#>  -2.9544  6.2757  2.5701  2.1514  3.2499
#> 
#> (1,3,.,.) = 
#>   -0.8730   4.6261  -0.0713   1.5787  -0.5837
#>   -5.2227  11.4824   1.6261  -1.3115   1.5294
#>  -10.5193   7.5100  -1.4419  15.6323   4.4622
#>   -2.5458  10.4856  -7.5824   0.6744  -1.5769
#>    0.4021  -0.3476  -7.8556  -7.1887   0.4617
#> 
#> (1,4,.,.) = 
#>  -2.9737 -2.8825  2.2192  7.4221  2.5999
#>   0.0660  1.6638 -5.7005  1.8543 -0.8390
#>   3.6661 -7.9368  3.6147 -3.7090 -8.0075
#>  -3.8910  4.2207 -7.3066 -7.1782 -9.1958
#>   6.7856 -0.5800  0.1512 -0.1717  0.4105
#> 
#> (1,5,.,.) = 
#>   0.2765 -5.8126 -4.1142  4.3612 -3.7084
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]