Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    3.6881   7.5130   1.8931   7.6466  10.5382
#>   -6.2439  -5.0594   8.0276  -2.7152  -0.6413
#>   -0.6454  -5.1329  -3.4176   2.5416  -2.2762
#>    4.9315   1.3029   0.2305  11.8345  -1.2072
#>    1.9268 -12.3453  -8.6864   6.0352  -0.8010
#> 
#> (1,2,.,.) = 
#>  -4.5628 -9.1821 -1.8513 -6.7035 -4.0778
#>   3.2325 -0.8917 -5.7812  1.5156  1.6146
#>   4.8218  7.2218  1.6169  2.5388  5.7983
#>  -2.6048 -1.0106  0.9556  8.3184  2.4295
#>  -0.3476  4.8575  5.5803 -4.2982  2.1599
#> 
#> (1,3,.,.) = 
#>   -5.0727   2.5073   0.8033 -17.8695   2.9446
#>    6.1768   2.2241   2.9414   9.8695   2.9513
#>   11.9046  10.2635   3.2570   0.1367  -4.7675
#>    0.2129  -0.5106   0.9899  16.6762  -3.9936
#>    2.1696  -0.7637   7.7555  -5.6012  -2.0362
#> 
#> (1,4,.,.) = 
#>    7.2135   0.5507 -12.1521  -0.5234   3.2985
#>   11.0638   1.4605  -1.4582  -5.1213  -0.9797
#>    2.3236   2.1269   8.5266   2.1441  -3.8975
#>   -7.2622   6.7486   7.0013   2.2694  -4.6585
#>    0.1262  -1.1140  -3.4138   3.7583   2.0110
#> 
#> (1,5,.,.) = 
#>  -5.3583  0.4189  8.2891 -7.6843  5.7965
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]