Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   0.2714 -2.1727  4.3527  7.0700 -0.4855
#>   1.9246  2.0981  3.5143  4.5149  1.2158
#>  -0.7428  0.8931 -1.2024 -1.8881 -11.9442
#>   2.5194 -3.7886 -4.5521  5.8015 -5.3621
#>  -0.1353  2.4567 -0.6896 -1.6948 -1.0737
#> 
#> (1,2,.,.) = 
#>  -5.4841  2.3950  2.4314  2.4434  5.8382
#>   0.6805  1.3032 -2.6787 -0.9202  3.0685
#>   2.4105  8.0551  4.8365 -12.7317 -6.4618
#>  -2.6319  3.8812  4.3438 -7.2876  2.1860
#>   3.8451  2.1758 -4.4303 -11.7497 -9.1270
#> 
#> (1,3,.,.) = 
#>  -3.5884e+00 -6.4840e+00 -5.4368e+00 -8.3600e-01  2.8909e+00
#>  -9.4964e+00 -2.3730e+00  8.7326e-01  3.1874e+00  1.2503e+01
#>  -4.6550e-01 -1.2143e+01 -3.6401e+00 -2.9278e-04 -4.3956e+00
#>   5.2258e+00  9.9204e+00 -1.0448e+01 -1.3335e+01 -2.7914e+00
#>  -3.8734e+00  2.2624e+00  1.7429e+01 -2.2254e-01  1.1533e+00
#> 
#> (1,4,.,.) = 
#>  -0.9574  1.6925  4.7072  0.3610 -1.5213
#>  -0.6132 -3.9057  4.8874  9.4002  3.5746
#>  -3.3077 -5.0796 -0.4670 -3.4934  2.4177
#>   3.1791  7.5194  6.5913 -0.8040 -4.3955
#>   4.4913  3.2497  2.9451  8.6855 -2.7148
#> 
#> (1,5,.,.) = 
#>   2.7817  6.0581 -3.5419 -1.1938  2.4027
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]