Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.8857   4.6894  -1.8277  -8.3494  -3.4774
#>    1.8840   5.6972   5.2248   3.9043  -0.7260
#>    3.7971   1.5144   1.4888   1.1784  -2.4556
#>   -2.7086  -3.1413  -7.9678   7.5697  -7.6435
#>   13.8567   2.7401  -5.8747  -6.7544  -5.6037
#> 
#> (1,2,.,.) = 
#>   11.1816   2.1492  -7.2258   8.3113  -4.8727
#>   -9.5219  -8.1226 -13.1149  -4.4085  -5.5936
#>    2.3947   6.7703   5.6743  -6.6260   0.6238
#>  -22.5772  -7.8539  -5.6331   9.8076  -3.2706
#>    0.1706  -1.9204  -7.8771  -3.3768  -0.1477
#> 
#> (1,3,.,.) = 
#>  -3.2627 -8.6749 -2.5824  3.2618  2.5054
#>  -5.6837 -8.1080 -10.7818 -0.0189 -2.2993
#>   2.8006 -5.1464 -6.8338 -5.0547  1.9066
#>  -4.1981  1.7703 -0.6369  5.3172 -2.3868
#>   0.0861  1.5562 -0.1276  1.9282  3.5618
#> 
#> (1,4,.,.) = 
#>   1.5922 -2.1906 -5.2656  1.8393  0.3666
#>   4.2043 -0.1136  0.0819 -1.5510  1.0824
#>   1.5640  2.4739  4.1278 -1.7269  1.3727
#>  -7.9118 -6.3904  2.5528 -5.7731  2.1250
#>  -4.9770  1.7755  2.2979  4.1057  2.5546
#> 
#> (1,5,.,.) = 
#>  -0.6853  3.0464 -6.9957  2.8151  2.0752
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]