Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -7.6735  3.8218 -1.1268  0.5836 -7.7238
#>   4.3007 -2.6118  5.9236  4.4666 -8.7301
#>   0.0732 -6.7250  6.0607 -6.4496  4.1717
#>   8.7505 -10.5013 -7.1034 -3.4442 -6.0673
#>   0.3951 -2.9024  0.8262  0.9931  2.7984
#> 
#> (1,2,.,.) = 
#>   4.3549 -8.7734 -2.8873  6.2454 -6.3309
#>   8.1161  1.4680  2.6218 -8.8734  4.6381
#>  -2.9973  4.5181 -9.8937  4.5182 -7.8824
#>   1.4852  6.2117  5.1413  5.7890 -4.0188
#>   4.7392 -6.6432  6.3406 -1.2282 -3.5699
#> 
#> (1,3,.,.) = 
#>    5.0148  12.0248   4.7321  -0.8703  -2.7005
#>   -1.4772  -2.0356  -0.4417  -4.8807  -1.1177
#>    8.8902   0.0645  12.5370 -11.7971  -2.3371
#>    0.6848   7.8446   8.7878  -1.7068  -4.3426
#>   -4.3245   1.5569  -2.1220  -8.3576  -3.0089
#> 
#> (1,4,.,.) = 
#>    2.2636   6.2839 -11.1621   5.4778  -7.9533
#>    6.2587  10.1305   6.8937  -4.9380  -0.2374
#>   -0.0337   0.4094   2.1913   7.1412  -3.0071
#>   -5.6419  -7.2525   3.8448  -8.5877   8.3326
#>    2.6774  -0.2804   3.9134   1.7939  -1.0552
#> 
#> (1,5,.,.) = 
#>  -2.9964  2.8338 -3.6696  9.2225 -0.6604
#>  -4.2823 -2.1924  3.7768  8.2519  6.2391
#>  -5.7546 -7.2860  6.1309 -1.9675  3.2649
#>  -0.2387 -0.7773 -1.7191  4.2507  3.2679
#>   0.7838 -0.0120  2.0476 -1.9527  3.3518
#> 
#> (1,6,.,.) = 
#>  -2.0417  4.7352 -11.3467 -3.8213 -2.1507
#>   2.2568  7.0305  7.0946 -5.1165 -7.2052
#>  -1.0859 -8.5605 -1.6991 -5.5085 -1.9433
#>  -3.9826  3.2366 -3.1493  1.9023  7.1238
#>   9.2984 -9.2219 -3.8753 -5.7850  2.5558
#> 
#> (1,7,.,.) = 
#>    2.2984  -7.2708   2.5572  -0.4221   4.5783
#>    4.4205  -7.4531  -5.4683  -7.3526   3.4383
#>   10.3840   2.6110   3.5127  -2.4053   5.9192
#>   -4.7132  15.3764  -1.2534   2.9533  -3.5840
#>    3.5144   4.7119   5.3858  -0.2218  -5.0744
#> 
#> (1,8,.,.) = 
#>   2.5451 -9.6130  7.1983 -8.0562 -9.0621
#>   9.4329  4.0261  6.8114  1.7317 -4.0614
#>   3.0090 -8.8324  0.1302 -1.4396 -3.2245
#>   5.5244 -4.3607 -1.3194  0.6479  3.6777
#>  -3.7887  1.4224 -1.2965  9.5453 -1.5017
#> [ CPUFloatType{1,8,5,5} ]