Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    0.3132   1.2809   2.0454   3.6291  -5.5063
#>   -3.2703  -4.1764   3.1663   1.4489  -1.7373
#>   -0.6250  -0.2749   1.4866  -7.9303   0.6104
#>    3.8106   8.6607  11.7270   2.8937   7.1433
#>    5.2020   4.3328  -0.1712   3.1150  -0.9931
#> 
#> (1,2,.,.) = 
#>  -1.3097 -3.1904  4.3048  4.2824  2.9307
#>  -5.3960 -7.1453 -2.6562  5.5261 -4.2124
#>  -9.3776 -1.7007 -0.9522  3.3312  4.5610
#>  -13.0524 -0.5931 -2.4766 -7.1636  0.2224
#>   2.0989 -6.6119 -1.0951 -1.6980  0.9297
#> 
#> (1,3,.,.) = 
#>  -1.5664 -2.4280  7.4189 -4.1340 -0.5252
#>  -3.0967  1.5198  3.9271  7.7336 -6.7853
#>  -0.0221  7.8211  2.1636  2.7581  2.1590
#>  -7.1927 -1.5670  2.3505  0.6146 -4.1000
#>   4.9643 -4.9224  2.1986 -2.5239  2.1865
#> 
#> (1,4,.,.) = 
#>   -1.0182  -0.0195   3.4979   2.1108  -0.8196
#>    4.4061  -8.2369  -2.1553  -5.4122  -8.4372
#>    1.2672  -4.4612   0.1152  -0.9858  -3.9772
#>    0.6773   9.0328  12.9113  -0.0439  -2.2748
#>    1.8396   2.3442   1.8350  -3.9544   3.0127
#> 
#> (1,5,.,.) = 
#>    5.0290   0.6391   7.4680  -8.4396   6.5963
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]