Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   3.5037 -0.3191  2.1228  1.8545  2.2932
#>  -4.5958 -0.5254  3.9093 -4.3578 -0.0400
#>  -3.8129 -3.8532 -4.1982  0.1561  9.3478
#>   5.8564 -3.8032 -3.3646  6.9294  5.7806
#>   4.7936  3.5476 -2.7111 -9.3776  0.7279
#> 
#> (1,2,.,.) = 
#>  -0.7376  5.1048  3.0366 -0.8054 -1.7234
#>  -2.1406 -3.3960  0.3679  4.2236  3.1871
#>   7.8214 -6.7899  2.2308  5.8667 -2.9255
#>  -4.1499 -1.9848  0.9015 -4.5280 -1.4574
#>  -4.8481  1.0109  3.1580  1.2861  2.3159
#> 
#> (1,3,.,.) = 
#>    0.2807   2.3917  -0.5172  -4.2956   3.2269
#>   -2.1749   4.8624  -6.5668  -0.9357   8.3202
#>    3.7172  -6.4782 -11.8935  10.7957   1.2051
#>    2.3176  -2.4627  10.9527   6.5446  -0.6262
#>   -1.8927   0.8731  -5.2645   5.2372 -12.6144
#> 
#> (1,4,.,.) = 
#>   -5.8356   9.0277   2.8029   0.8123  -3.9050
#>    3.0282  -2.0011 -10.8925  -0.1510   2.7547
#>   14.0584  -1.1806   1.7930   8.7237   4.8085
#>    2.0803   3.9829  -9.3425  -4.1649   4.7704
#>    0.4661   6.5541   1.2623   4.0593   5.2751
#> 
#> (1,5,.,.) = 
#>    4.7161  -4.6995  -6.4484   7.0897   4.0255
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]