Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    2.4109  -4.2655 -12.6756   2.5068  -5.3319
#>   -6.6049  22.3909  -4.4156   9.2278  -5.8447
#>   -1.4485   7.2239  10.9439  12.4157  -0.7311
#>   -2.0877   0.8160   4.3186  -0.1174   1.6761
#>    1.3799  -2.3042   1.6199  -8.4368  -1.2032
#> 
#> (1,2,.,.) = 
#>   2.0096 -9.9029  4.5905  0.4468  0.7658
#>   4.5052  5.7395  7.9990  5.6396 -1.4025
#>   1.9102  4.9814  7.0757 -2.5934  4.2325
#>   3.9160  4.7586 -7.1393 -5.5299 -0.3106
#>  -2.5669 -0.8903 -9.7714 -1.4656 -2.8818
#> 
#> (1,3,.,.) = 
#>   -2.1977  -0.5229  10.9863  -6.8118  -7.2412
#>   -0.5534  -6.4873   3.6370   8.1065   5.0024
#>   -3.7277   5.2004   1.1771  12.2045  -2.8321
#>    3.9320   1.1555  -1.5095  -4.4209  -6.4210
#>   -1.1893  10.6942   6.8042  -2.3287   1.2248
#> 
#> (1,4,.,.) = 
#>  -0.1279  4.7366  7.6807 -0.5141 -0.1526
#>   0.1436 -0.0779 -4.6899  4.0344  5.6323
#>  -6.0219  4.9467 -4.6613 -1.7882  7.5360
#>   2.1438  1.6196 -11.1295 -0.2598 -1.2540
#>   2.8289  1.5394 -3.1957 -7.2217 -2.3609
#> 
#> (1,5,.,.) = 
#>    0.3477  -1.0651  10.8236  -3.6129   7.3720
#>    1.4985  -0.7103  -5.1579   4.9092   4.3412
#>    2.9074  -0.3843  -5.1243  -7.4555   2.8379
#>    3.4378   3.9050  -3.4318  -1.4930  -0.5527
#>    1.1783   4.7545  -0.4864   3.8900  -4.2654
#> 
#> (1,6,.,.) = 
#>   -3.1013   4.4041  16.3512   4.4981   0.1130
#>    7.9521   0.3651   7.3655   9.8481   0.9426
#>   -4.3156   4.2464  -2.9211  -0.6906  -1.0960
#>    2.5792  -1.6355  -7.6319   3.4977  -0.0904
#>   -0.8454   4.0276   0.3242  -6.0821  -2.7442
#> 
#> (1,7,.,.) = 
#>   -0.5626  -8.0822   7.4913   3.7802   3.2062
#>    2.1506   7.7184  -4.9659   0.5388   4.2373
#>    0.1397   5.2730  -8.5830  13.3731  -3.8212
#>    1.1913   1.6860   9.0611   4.6355   0.3181
#>   -8.7467   3.2245  -0.4988   9.7048  -0.9619
#> 
#> (1,8,.,.) = 
#>   -5.3833  -3.2286   6.4152  -3.3700   2.0266
#>    0.2035   2.6462   2.1349  -6.5291   1.9261
#>    0.1194  -3.8128  -2.8644  -8.4293  -1.9006
#>    5.5594   2.2421  -3.1574  -6.1346   0.3109
#>    1.2211  -2.3825  -6.1089  13.0139   4.6268
#> [ CPUFloatType{1,8,5,5} ]