Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -1.6205 -10.3339  -2.0657  -4.0155  -3.6359
#>   -3.0703  -4.9796  -3.6777   5.0390  -4.4537
#>   -1.6369  -3.1138  -0.0413   5.9534   1.2015
#>    9.4982   3.8842   0.3703   1.7477   0.5988
#>    3.4309   5.2623   0.8649  10.9743  -6.3994
#> 
#> (1,2,.,.) = 
#>    4.0828   1.8423   0.1880   3.4531  -1.6986
#>   -6.3644  13.4195  -1.2169   8.7391   0.0999
#>    0.0282  14.5723  -0.4932   5.7594   3.4258
#>   10.0407   4.9981  -2.2772   2.1945  -4.6350
#>   -0.2642   3.9606  -4.4851   4.5449   0.1136
#> 
#> (1,3,.,.) = 
#>   0.1062 -0.6615  2.6222  8.5964  6.4844
#>   9.1932  7.9341  7.3005  2.6670 -4.0014
#>  -1.2468  2.0834 -8.7934  1.9467  2.0677
#>  -11.2430 -7.9170 -3.6361  1.9448 -0.2099
#>   1.7563 -5.5882  6.2709 -3.0580 -4.9351
#> 
#> (1,4,.,.) = 
#>  -2.5014 -5.9059  0.0595 -3.5559 -0.9859
#>  -1.2308 -1.8009 -4.9836 -7.8489  6.0853
#>  -4.1321 -0.1922 -6.8452 -4.9848 -0.6688
#>  -1.6822 -6.9856 -1.3200 -6.6738  4.0593
#>   1.6590 -2.6009 -0.7329 -7.6106  1.7449
#> 
#> (1,5,.,.) = 
#>   -2.1775   9.0733  -1.4800   4.4329  -1.5106
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]