Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -0.0622 -1.3519  4.6962 -3.5689  0.0513
#>  -6.8808  2.8543  0.5661  5.5569  3.8458
#>  -4.9593  6.1840 -0.2277  6.3055  0.1379
#>   7.6113 -0.6232 -0.6824  1.0828  6.0511
#>   5.6334  0.0086  2.1451  1.7260  1.9646
#> 
#> (1,2,.,.) = 
#>   4.2390 -0.8827 -3.1172  1.1444 -2.0940
#>   5.6135  2.5332  1.8957 -9.7463  1.6421
#>  -7.3771  3.6384  3.2471  0.3165  4.1442
#>  -4.0548  9.0570  6.3252 -2.6788  2.5611
#>  -6.9180 -8.6890  3.7634  5.9058 -3.6733
#> 
#> (1,3,.,.) = 
#>  -0.4894 -2.6945 -2.1639 -0.2275 -1.2641
#>  -5.5115  8.7296  8.3885 -2.1748 -0.9266
#>  -11.3998  0.4643  6.7172  4.3789 -2.4128
#>  -1.5630 -9.1956 -11.1178 -4.3789  4.0429
#>   8.7368  4.5744  5.6954  2.5964  0.2247
#> 
#> (1,4,.,.) = 
#>    4.2095   1.2731  -3.1387   8.4913   0.2902
#>    4.6789   6.1968  -9.3846  -7.3187  -4.6580
#>    0.0241  13.7853   7.6048   2.5640  -2.4263
#>    1.4961  -5.6421  -2.4943  -0.8850  -1.5088
#>    0.5919   0.2364  -6.5464   1.7723  -6.2718
#> 
#> (1,5,.,.) = 
#>   3.8337  2.9579 -2.6305  4.2065 -0.1964
#>   7.6632 -5.4463 -2.5970 -7.5247  2.0145
#>  -1.7723  2.3882 -0.5465 -1.5220 -5.2836
#>  -4.3007  0.7113 -4.3532  1.0323  0.6968
#>   3.8143 -6.9671 -0.0214  2.0123 -3.8443
#> 
#> (1,6,.,.) = 
#>  -1.3959  7.4730  5.2863 -0.1943 -3.7167
#>  -5.2690  6.9239 -0.2235  7.2675 -1.3668
#>  -0.3263 -14.4025 -5.2340 -4.3457 -0.0798
#>   9.4260 -9.9315  6.3413 -0.8455 -1.2604
#>   3.9679 -8.0024 -7.5344  1.9896  3.6740
#> 
#> (1,7,.,.) = 
#>   -2.8713   0.8366  -0.1888   5.8816   5.7257
#>   -0.2640  13.7833  -5.5540  -5.3195  -2.6448
#>  -12.1150  -4.6795  -4.2531  11.2328  -3.9452
#>    2.9817  -3.0572 -11.1085  -2.0123  -0.4154
#>    2.6884   0.1765  -1.1556  -1.2219  -1.4902
#> 
#> (1,8,.,.) = 
#>  -9.6729 -4.4364 -9.7855  0.4185  5.1251
#>  -1.2360  3.0474  5.6872 -1.7649  2.7610
#>  -5.9558 -0.2239  1.7186  6.3789 -1.8982
#>   3.5773  8.7840  2.5906  2.5301 -3.7890
#>  -3.9069  0.5540  5.1767 -0.5805 -1.5124
#> [ CPUFloatType{1,8,5,5} ]