Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   1.2677 -0.3564  0.6463  2.5518  0.5761
#>  -5.1650 -0.9972 -7.4322  2.2633 -2.4653
#>   3.0212 -0.8861 -0.5157 -0.6335 -1.5622
#>   0.4730 -3.6966  2.0442 -1.1706  0.7480
#>  -6.8055 -7.3875  0.3424 -0.9073  1.3198
#> 
#> (1,2,.,.) = 
#>   1.6691 -0.2127  1.2938 -3.5512 -2.5449
#>  -2.4997  6.1252 -5.4271  3.7257 -3.3927
#>  -0.6673 -1.8647  1.5452  1.7572 -6.1349
#>  -6.8039 -0.5486 -1.6320 -3.7800 -3.1728
#>   2.6114 -3.0218 -7.5187 -2.0001 -3.5111
#> 
#> (1,3,.,.) = 
#>   -1.2270  -0.9526   0.1490  -1.3736   3.0288
#>   -1.2634  -2.2485  -4.8966  -2.6926  -1.6630
#>   -7.2981  11.6269  -0.1175   0.4318  -3.5438
#>   -1.4742  -5.8229  -3.5807   0.2386   0.9544
#>   -0.2849  -4.2705  -2.6006  -3.3688  -5.4811
#> 
#> (1,4,.,.) = 
#>   6.8087 -2.7165  3.9721 -0.4758 -3.1887
#>  -3.6976  1.5788  1.6681 -2.9425 -4.5742
#>   2.9259 -7.1392  1.8834  1.4178  2.2408
#>  -10.3528 -0.7246  4.0790  2.9147 -1.6664
#>   7.0298 -2.1101  1.0677  3.1038 -1.8439
#> 
#> (1,5,.,.) = 
#>   2.5335  0.9699 -0.4180 -0.2923  0.0059
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]