Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.6821   2.7374   8.7786   3.0126  -0.9494
#>    7.7186  13.5661  -3.7085  -5.6252   2.4869
#>    4.4679   1.9995  -9.3365  -0.6740  -1.3210
#>   -0.0523   1.5059 -12.6984   0.6196   5.7581
#>    4.9991  -1.1776  -0.1596   1.4679   0.3769
#> 
#> (1,2,.,.) = 
#>    3.9535  -6.1813  -0.0470  -1.1963   3.3869
#>   -0.5509   4.7392  -4.8215  11.9052   0.1826
#>   -9.2325   9.2761   1.0280   4.3329   6.2150
#>   18.1562   1.2495  -1.2615  -5.3601  -5.2685
#>   -8.0194   2.1725   3.1715  -1.5496  -2.5495
#> 
#> (1,3,.,.) = 
#>    8.5860  -4.7394  -3.1089  -3.7410   0.9926
#>   -1.5004  -2.1219   0.1617   7.6442   1.6011
#>   10.9793   9.3642   3.4885  -8.6970  -4.6602
#>   -5.2647   1.3758  -3.1785  -1.2251   1.3556
#>    3.7179  -2.2696   3.0935   3.2901  -0.9679
#> 
#> (1,4,.,.) = 
#>   2.6082  6.3329 -3.0712 -2.8107  0.3301
#>  -3.8301 -3.8183 -2.2991  5.4530  8.6846
#>   2.1420 -5.0257  4.0553 -5.0042 -2.0850
#>  -4.7691  6.3787 -7.1825 -5.7990 -4.8059
#>  -3.9639 -2.8912  1.9015 -10.6152  9.0753
#> 
#> (1,5,.,.) = 
#>    5.3456   1.7697   4.9987   2.2996   2.2223
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]