Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -3.4066   5.5283  -5.0377   0.6698  -4.4284
#>   -3.6934  -2.4781   2.5744  -3.0031   0.4655
#>    3.6889  -6.9205   5.3310   4.3698  -1.2455
#>   -2.0307  -4.2932  -7.4053  -5.8052   0.3443
#>    1.4582  -2.9901  10.3897   7.2761   1.1578
#> 
#> (1,2,.,.) = 
#>    2.1803   0.3849  -1.5164   1.2051   3.4360
#>   -1.3907   6.0862   1.5743  -3.4517   4.1087
#>    4.9837   3.9837  -0.2061   3.7555   0.8809
#>    2.2718 -11.2095  17.3016   4.5389  -1.7334
#>   -6.5350   2.4432   1.3544  -2.3759  -0.5317
#> 
#> (1,3,.,.) = 
#>    5.1603   6.8404   2.3208  -1.1607  -2.2306
#>   -5.9020   2.6310   8.3586 -12.1010  -4.1656
#>    0.0820  -0.4396  10.5927   3.9376   1.2618
#>   -7.4381  -4.0070  -4.3925  -3.4717   2.3439
#>    6.2061   5.4509   4.0905   3.8469   5.3138
#> 
#> (1,4,.,.) = 
#>    1.0866   7.5523  -0.8344  -7.6081   9.3892
#>   -5.9134   5.4492  11.5496  -2.1199   5.1301
#>   -4.1519 -10.8839   0.4673  -3.5534  -6.5269
#>    7.7621  -6.6516   9.3200   9.2781  -1.1712
#>   -4.5972   4.8612  -7.0155   1.5001   2.3273
#> 
#> (1,5,.,.) = 
#>    4.5623  -1.3311  -0.9767  -0.6893   1.4017
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]