Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   0.5821 -2.0186 -0.8549 -12.7129 -1.2240
#>   0.4415 -5.4739 -8.9374  8.7621 -2.4748
#>   0.3764 -4.6335  1.6122 -3.3858  3.6876
#>   4.3834  1.2873 -1.8663 -0.3002  0.2717
#>  -0.6433 -9.7659  0.2099 -5.9349 -1.3401
#> 
#> (1,2,.,.) = 
#>  -4.1171 -4.1934  0.2312  8.9871  4.1412
#>   5.9788  2.3210  6.5586  1.2520  5.0636
#>  -2.4891  7.2192  4.1434 -0.6973  0.6995
#>  -4.9581 -9.8072  4.3166 -7.2285 -10.3275
#>   0.8496  8.3685 -5.0551 -0.4267 -0.3471
#> 
#> (1,3,.,.) = 
#>   0.3061  4.4260 -7.0666 -8.7521 -4.0431
#>  -3.0963 -2.6705 -1.8009  7.0072 -0.6566
#>   2.6096 -0.0878  2.7711 -5.3469  2.5064
#>  -0.4161  8.5284 -3.0559 -8.5501  2.6072
#>  -1.4028 -9.0858 -2.3916  9.5749 -0.3775
#> 
#> (1,4,.,.) = 
#>   1.3558  7.5624  1.5048  5.5938  1.2264
#>   0.9392 -1.3030 -2.8354 -5.8062 -3.3344
#>  -0.7884 -0.4168  5.9915 -1.5218  4.7651
#>   0.2086 -4.5958  2.0162  1.7478  0.0038
#>  -5.0495  0.5539 -9.9076 -3.8326 -1.8840
#> 
#> (1,5,.,.) = 
#>    4.7900  -5.1779  -1.2105  -9.4398   8.7043
#>   -0.1843   1.8321  -1.1734  13.5041  -3.4905
#>    1.1003  -3.1419  -0.3899  -8.6008  -3.2421
#>   -5.4268   5.4738 -10.0630   7.5114  -5.2629
#>    2.5506   3.4341  -0.9392   2.5237  -5.2635
#> 
#> (1,6,.,.) = 
#>   -1.6692 -10.3936   2.7194   0.5011   2.8448
#>    4.8521   0.9378   3.6040   7.3668  -7.1894
#>    0.8318   4.2670  -3.0975   0.7498  -3.7483
#>   -1.6722   0.5094   5.9559  10.8624   8.5367
#>   -0.6876   2.4163   3.5673   0.1746   3.6481
#> 
#> (1,7,.,.) = 
#>   -0.7487   0.7485   4.0581  -1.8421   0.5531
#>    0.7256  -1.6724   7.4091  -5.2383  -0.8266
#>    0.2720 -12.7966  -6.7527  -1.4524   5.3720
#>    8.2702 -12.2665   9.7300   0.7959   1.0094
#>    1.4836   0.9964  10.8320  -5.1627   1.7035
#> 
#> (1,8,.,.) = 
#>   0.9140 -4.7645 -4.9583  4.7891  3.7165
#>  -6.7059  6.6868 -3.7802  7.0551 -5.6355
#>  -1.3517  1.0797  8.0352 -3.2764  4.4821
#>  -3.1988  0.3283 -13.1845  2.3922 -5.9737
#>  -8.2763 -1.0262 -3.2317 -7.8120 -1.2987
#> [ CPUFloatType{1,8,5,5} ]