Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    0.9430  -5.4281  -7.0785   2.8299  -2.1493
#>    3.4462  -3.8896  -3.2722   6.4524 -10.4976
#>    6.2292  -3.6948  -4.1844  18.1900  -9.0877
#>   -2.8189   5.2771  -4.3308   5.3131   1.2688
#>   -0.4264  -3.0181  -5.8961   1.4272   1.6727
#> 
#> (1,2,.,.) = 
#>   -0.7695  -5.5408  -4.6155  -5.4872   1.5503
#>   -1.8370  -7.0362  -4.3154  -2.8237  -1.7803
#>   -0.0375   5.3829   4.6701  11.7793  -0.0717
#>   -2.4118   4.0060  -8.4304  -3.6248  -4.2981
#>    1.2284  -5.0404  -0.2036  -1.2800   2.3839
#> 
#> (1,3,.,.) = 
#>    0.9262   4.7282  -5.2759   9.3573   1.9843
#>   -1.3711   4.9418  -2.2793  -6.7944  -3.9839
#>    4.9475   1.1510  -2.3306   4.1822  -1.8550
#>   -0.5368  19.1446  -2.1089 -11.1088  -8.0311
#>    4.4127  -9.3071  10.9435   0.3096   5.1700
#> 
#> (1,4,.,.) = 
#>   0.7515 -1.7027 -3.1393  3.9436  3.1533
#>   5.9778 -2.9573 -3.2372 -4.9740  0.2948
#>  -0.9628 -8.2274  2.8003 -0.0742  1.9083
#>  -1.5258  6.3115 -5.4333  6.2672 -6.6831
#>   0.1119 -2.8011  4.5451  6.2484  3.8557
#> 
#> (1,5,.,.) = 
#>   1.9085 -5.2575 -5.0556 -1.5399 -5.9149
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]