Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -0.7928   0.5078  -1.3043   2.5885  -2.6018
#>   12.5521  -3.7497  -0.6191  -1.6196   2.3922
#>    0.0056   8.8994  -5.4218   8.3781   3.9310
#>    9.2635   8.5440   2.0032   2.4436 -11.3406
#>    3.8988  -2.1977   8.4364   9.6339   7.8368
#> 
#> (1,2,.,.) = 
#>    3.0357  -0.1387   2.0027   9.8047   0.8293
#>   -8.2204   1.6309  10.0764   2.9889   1.1089
#>   -4.4243   3.7857  -7.0363  13.2038  -2.9072
#>    1.4941  -3.1273  -1.5765   0.0046   2.0442
#>    3.1726   0.1137  -0.6356  -0.6416  -2.9584
#> 
#> (1,3,.,.) = 
#>    5.7665   4.0157  -3.2663  10.7996  -3.2159
#>   -6.1252   3.7690   2.0231   2.1532  -0.6732
#>    6.5913  -1.2416  -3.9490   1.0089   0.4850
#>   -4.4972  -5.2271 -10.5996   4.7195  -1.5154
#>    1.4491   3.9479  -0.4126  -8.4575  -3.2938
#> 
#> (1,4,.,.) = 
#>   -1.6933   1.4590   0.4266   6.9579   4.4207
#>    0.3553  -0.7029  11.0479  -7.7245   3.4540
#>    5.9933   6.5318  -1.1687  -1.2272  -3.9242
#>    1.1664   5.6510   3.5530  -3.6595   7.8148
#>    2.2475   1.9327   3.8911  -2.3189  -5.0795
#> 
#> (1,5,.,.) = 
#>    0.0234  -1.3201   1.8387  10.3785  -3.7975
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]