Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   1.7751  1.8102 -5.3800 -3.0488 -5.4477
#>  -5.7712  2.6865 -0.0846  7.5752  6.9191
#>   5.2896  3.0661  9.7657 -15.9121 -0.4981
#>  -9.0795 -1.5815  3.5357  9.3267  4.3401
#>   2.0030  4.1984 -14.2355  0.2379  5.6598
#> 
#> (1,2,.,.) = 
#>    3.7265  -0.2746  -1.2940  -1.8794   3.8898
#>   -1.3463   4.1611  11.7422   0.3257  -2.8260
#>   -3.1741  -9.4664   0.3786  -9.1628   2.2345
#>   -3.9877   8.6068   3.5048   4.0214   3.5298
#>    1.4579   0.8858   0.2262   0.5585   1.9909
#> 
#> (1,3,.,.) = 
#>   2.8495  7.5995  2.6102  4.4934 -4.1878
#>  -4.5630  3.2005  4.1013 -2.8311  2.1538
#>   3.3536  5.8931 -4.5269  4.2960 -4.2912
#>  -10.8705 -0.3137  6.4640  0.5994  6.9229
#>   7.0797 -0.3412 -1.9438  4.4663 -0.2123
#> 
#> (1,4,.,.) = 
#>  -1.0945 -0.2992  5.3942 -4.6463 -3.2643
#>   4.5450  0.4313  2.9448  8.1186  6.1091
#>  -5.4938  8.1003  0.6197  1.0535 -8.3222
#>  -6.4377 -8.4517  6.1147 -1.4768  1.7856
#>   0.4933  1.0665  3.6550 -1.2144 -6.8495
#> 
#> (1,5,.,.) = 
#>    1.9912   6.3521  -6.3074   5.0286   1.5800
#>    4.6302  -6.5545   1.2782  -9.8425   1.6922
#>   -1.9579  11.3642  -0.7866   2.6508   2.1056
#>   -4.8874   0.3786   3.4781  -8.8948  -6.2078
#>   11.8482   3.8567  -0.3391  -0.4456   5.6384
#> 
#> (1,6,.,.) = 
#>    3.2679   1.4055  -6.6733   9.1439  -0.1774
#>   -3.6459   1.0924  -3.4128  -9.4440   6.2553
#>    1.4464  -5.3091  11.0492  -4.1211  -7.3963
#>    7.6827   7.8154  -6.3949   2.2058  -7.1125
#>    0.1458   9.6769   3.2690   1.7532   0.0174
#> 
#> (1,7,.,.) = 
#>   -5.2135  13.6151  -3.6709   2.1498  -3.5713
#>   -0.4493  -4.8700   3.7786   2.1324   8.0006
#>    4.3950  -5.6337   7.7445  -8.1921   0.6682
#>  -12.5536   2.6553  -3.4464   6.5250  -5.0772
#>    0.1851   6.6151   1.0536  -3.3529   1.2139
#> 
#> (1,8,.,.) = 
#>    0.3621   2.2075   3.1951   1.3277  -2.4578
#>    5.0692   2.1762  13.6334  -4.1285  10.3525
#>    7.2175  -5.1844  -1.5165   7.2140  -6.2546
#>   -8.7801 -13.0705  -3.9751  14.0640  -9.8164
#>   -2.6648   1.6580   4.6547   0.6180  -0.3127
#> [ CPUFloatType{1,8,5,5} ]