Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -5.7637   0.7574   1.1417  -6.5483  -4.6139
#>    1.4272  -5.3505  10.6444   5.5477  -0.6306
#>    1.5011  -3.6194  -3.9702  11.1942   2.9240
#>   -7.6165   2.7631  -0.8415   8.6983  -3.8943
#>    2.5300  -2.2918   2.9979  -5.8671   8.8397
#> 
#> (1,2,.,.) = 
#>  -3.1471  0.8954 -0.5328 -2.2513 -3.4303
#>   6.5634 -7.9835 -0.0346 -0.1767  2.8531
#>  -7.5841  4.1301 -8.9242  8.3177 -0.3384
#>   0.4684 -7.6158  3.5138  1.3095 -3.6137
#>   6.6998  9.7681 -4.8432 -5.5394 -2.5476
#> 
#> (1,3,.,.) = 
#>  -5.7695 -2.4566 -0.5493 -7.6069 -4.2084
#>   5.5918 -7.9663 -0.1881  4.2266  0.5691
#>  -7.2210  2.2020  3.9159  6.5547  0.5018
#>   0.5902 -2.9667  9.5564 -9.1325 -1.7068
#>   1.5258  2.0763 -2.5673 -10.5571  1.3391
#> 
#> (1,4,.,.) = 
#>    3.1946  -1.4580  -2.3480  -0.2003   1.4259
#>   -7.2968  -1.6798  -7.9296  -4.9708   0.9479
#>   -0.5489  -1.3728  10.4354   4.7905   1.7829
#>   -0.2926   1.3479  -2.4394  -5.0893   6.3091
#>    0.7088   0.1754  -2.7283   2.2440  -1.7656
#> 
#> (1,5,.,.) = 
#>    3.6293  -3.5085  -1.1953   8.1782  -1.4827
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]