Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -4.0377  -4.3610   3.3060  -4.5761  -1.3726
#>   -2.2389   3.1566   2.4350   3.0903   2.8771
#>   -0.8279  -9.1534  -2.9221  -2.3449  -3.9698
#>    0.1888  -0.3639  -6.4356  -8.9568  -2.3284
#>    4.3070   2.6646   5.8210  11.8437   4.7287
#> 
#> (1,2,.,.) = 
#>   0.7242  1.1063 -1.6384  0.5639  4.0082
#>  -1.3614  1.7085 -1.2651  8.9471 -6.2028
#>  -2.4456  6.4605  6.0586  3.8198 -3.6963
#>   1.1772  2.6906 -8.9246  1.6811 -8.0847
#>  -2.6511  1.1572 -2.5389 -1.1193 -5.7530
#> 
#> (1,3,.,.) = 
#>  -1.0874 -3.4295  7.2229 -6.4289 -4.1342
#>   0.7638 -7.9233  6.6774 -8.4771 -1.6152
#>  -4.6208 -0.3764 -1.5297  7.8265 -0.8093
#>  -4.7092 -9.5604  5.9099 -0.0211  1.1666
#>  -2.9517  4.1237  7.3893 -3.4579 -0.1294
#> 
#> (1,4,.,.) = 
#>  -0.2744  0.5214 -4.3592  7.1092  1.7757
#>  -1.5699  5.3052  6.5393 -0.7288  0.1444
#>   7.1937 -0.6196 -8.5484 -6.7766 -1.6662
#>   4.2730  0.5406 -7.6268 -2.5635  2.8401
#>   4.9776 -0.0668  9.0504  5.5438 -2.0487
#> 
#> (1,5,.,.) = 
#>   -0.2167   0.8077   1.7214   1.4352   1.9897
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]