Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    1.1466   8.3791  -5.1339  -2.9047  -2.4335
#>    5.4468   8.6050  16.3585   7.8657  -6.1840
#>    5.5737   2.2005   7.7583  -6.4431  -8.2943
#>    2.7088   2.3053   0.8020  -1.0082   3.8738
#>   -0.4331   0.1507 -13.6889  -8.9055  -0.6310
#> 
#> (1,2,.,.) = 
#>   6.7149  1.8742  4.1062 -0.3791  2.3364
#>   3.3976 -4.4588 -5.5360  5.2566  0.2396
#>   4.3126  4.1048  1.9514 -4.9387 -3.5809
#>   0.8287  5.0659 -7.6905 -9.2746  2.2166
#>   3.3904 -1.9668 -0.5273  1.0774 -1.3944
#> 
#> (1,3,.,.) = 
#>   -0.1103   1.2774   7.3607   4.4346  -5.4044
#>    2.2745 -10.9541 -10.3554   0.7713  -1.4058
#>    0.1480   1.8984  -1.6501   0.9580   2.7086
#>   -1.7366  -2.4097   5.5815  -7.3540  -0.3734
#>   -0.5252   9.7074   3.4123  10.5416   5.4732
#> 
#> (1,4,.,.) = 
#>   6.7885  8.7417 -5.0142 -5.3350  8.3765
#>   7.5070  6.6185  5.5466  8.8818 -1.3335
#>  -0.1803 -2.6054 -1.7047  2.6237 -5.0197
#>   0.0180 -0.3135 -5.1092  2.6217  1.5574
#>   0.5320 -1.9768  1.3318  0.9128 -0.7111
#> 
#> (1,5,.,.) = 
#>    2.1363   0.7132  15.4153   1.4850  -6.7184
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]