Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -6.0379  0.4194 -2.8652 -2.2436 -0.0630
#>   4.7363  2.7267  4.3768  3.5500 -2.6843
#>  -0.6462 -4.3789  3.0130 -1.0171 -1.7568
#>   0.5392 -3.2822  1.9279  5.5512  3.0696
#>   3.1603  7.9060 -0.8370 -6.4351 -2.0993
#> 
#> (1,2,.,.) = 
#>  -0.0295 -0.9298  5.1778  5.6393 -2.2497
#>   2.3261  1.1071  1.7943 -2.6955 -4.5381
#>  -6.9073 -1.6443  1.7272 -2.5201  1.5468
#>   0.2224  1.6643 -3.5661 -8.2282  1.4448
#>   0.3579  0.1060 -13.2866  0.2225  5.7793
#> 
#> (1,3,.,.) = 
#>    8.2682  -2.9437   4.0937  13.3131   3.0799
#>    1.1723   0.7031  -3.5217  -7.5859   7.1652
#>   -2.3856   2.1407   3.9189   5.5789  -0.1286
#>    3.5301   6.5983   6.3807   6.3466   6.6055
#>   -7.5158  -4.1064  -1.2402   1.7808   1.5438
#> 
#> (1,4,.,.) = 
#>   -1.4731   4.8450   0.5016   2.7951   3.9236
#>   -5.0995 -15.2092   9.5128  -5.6531  -6.5906
#>   -0.4927   0.9667  -1.5701  -1.5797  16.3502
#>   -3.9475   8.7248   7.4626  -0.0081   0.0630
#>    2.4504   0.7484  -9.5465  -2.3768   6.1024
#> 
#> (1,5,.,.) = 
#>   -7.9878  -5.0489  -2.4314  -3.9968   2.2191
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]