Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   -0.0905   2.3159   5.1436  -6.0822   1.1803
#>    1.5350  -2.2645   8.2221  -9.9417   7.2463
#>   -7.7977  13.5817   2.6384   4.6236   3.8218
#>   -4.9535   2.3802  -0.8261   2.2480   6.4480
#>    4.0567  -1.6355  -2.0388  -1.4397   4.2229
#> 
#> (1,2,.,.) = 
#>  -1.0567  1.6206 -11.4827  5.5172  3.7397
#>  -7.3183  3.1484 -3.4809  0.1997 -1.1304
#>   4.7954 -12.7745 -2.6562 -10.2770  4.1020
#>  -0.3908  3.4519 -7.2051  3.9836 -2.0018
#>   2.2743 -7.6631  4.8773 -8.6464 -3.6239
#> 
#> (1,3,.,.) = 
#>   -6.4278   2.6778   3.4233   5.0826   0.3418
#>   -5.5157   8.1709   3.9151   7.6825  -5.4678
#>   -1.1152  -3.2462  17.1005  -7.5965  -0.8358
#>    3.1078   0.1476   1.2988  -0.0677   0.8684
#>   -3.0605  -5.3267   0.2946   7.6815   3.6938
#> 
#> (1,4,.,.) = 
#>   1.3705  5.6027  4.2027  3.3589  1.2279
#>  -1.1207 -0.3036  0.0281  9.1358  2.2815
#>  -3.6459 -5.4493  8.0393 -1.2888  5.3232
#>  -2.6161 -1.2935  0.0991  4.3635  2.1710
#>   0.5488 -6.7206  3.8062 -0.6501  0.0810
#> 
#> (1,5,.,.) = 
#>  -11.0002  -7.0254  -1.6860   9.7165  -1.1289
#>   -4.3386 -12.2846   9.3417   1.1901   0.5854
#>   -0.9123  -9.9780  -2.0725  -8.6009  -0.5795
#>    1.3227   5.1593  -6.8049  -1.7243  -4.2645
#>   -0.5701  -6.3302   0.4875  -4.8259   0.1266
#> 
#> (1,6,.,.) = 
#>   -1.2253   2.3183   0.4377   4.3394  -0.7717
#>   -2.2938  -1.4362  17.3881  -7.1605   4.4037
#>  -10.8767   8.0605   0.4145  -0.7609   1.2350
#>   -4.7865   5.3891  -3.3427   6.1662   5.5622
#>    2.4586  -1.4331  -3.0880  -4.7222   5.6322
#> 
#> (1,7,.,.) = 
#>    3.4931   5.2776  -1.8028  -9.2519  -4.2092
#>    3.5915  -4.0971 -10.0300   2.8863   6.4705
#>    4.6912  14.4797 -16.9482  -3.6167   0.0637
#>    5.6531  -6.6439   4.4269   6.9572  -1.7879
#>   -0.1376   5.0179  -0.6261  -1.6148   2.2901
#> 
#> (1,8,.,.) = 
#>   1.8480 -11.4047 -8.1572  4.0602  3.6089
#>  -3.8658 -7.0709 -5.1871  1.3399  3.8792
#>  -12.0865  3.2986 -17.2404  8.2520 -0.6314
#>  -0.8320  5.8538 -1.8970 -2.7349 -5.9085
#>  -0.6356  2.8053 -5.3046 -2.0396 -6.7148
#> [ CPUFloatType{1,8,5,5} ]