Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -0.5858 -4.0843  3.4392 -4.1538  2.4232
#>   5.8466  5.3796 -5.2865 -5.2637  1.9978
#>   0.5524 -3.9714 -6.1022  4.1460  3.8923
#>  -2.8976 -0.0619 -6.5124 -0.6081  6.0023
#>   2.3829  4.8095 -3.4515 -2.2901  2.4507
#> 
#> (1,2,.,.) = 
#>  -2.9948  3.3939 -2.5289 -10.5149 -5.4079
#>   1.1518  5.4048 -6.1932  3.4703  0.9714
#>   4.6111  2.4101  5.4068  3.8346 -0.7034
#>  -6.7358 -1.0116  2.1359 -4.7764 -9.4927
#>  -3.6616 -5.4944  2.2788  2.7387  6.3885
#> 
#> (1,3,.,.) = 
#>    3.9640  -0.6900   3.4976   6.1465   7.0068
#>    3.8170  12.3384  12.4775  -1.6670   6.2386
#>   -4.1021  -1.7235 -11.6157  -6.8295   6.2674
#>   -0.3119 -10.5032  -8.3573   5.3670   9.1851
#>   -9.0517   1.7361   6.6258   5.1008   2.4393
#> 
#> (1,4,.,.) = 
#>   3.3879 -7.3657  3.5964 -1.6155 -3.5032
#>   2.9438  0.2152  7.8949 -10.2920  0.7234
#>  -5.8244  3.3482 -2.2886 -4.3091  0.7224
#>   8.4309  3.3978 -1.6483  4.3497 -6.1477
#>   0.3542  3.2067 -2.1077  7.5392  2.9555
#> 
#> (1,5,.,.) = 
#>    2.7335  -1.7427  -1.6915   7.7144   0.6153
#>   -2.1443   2.8212   0.1702  -5.8547  -8.0375
#>    0.5057  -6.8014  -8.0627   3.8882   4.8378
#>    4.8681  -1.1328  18.7384   7.1581  -1.4027
#>   -4.8924  -2.8099   0.0269  -2.6088 -11.5160
#> 
#> (1,6,.,.) = 
#>  -6.6684 -3.2616 -1.0202  0.3536 -1.6571
#>  -6.1089 -4.6204 -4.2796  0.0391  6.4487
#>  -2.1313  6.6342 -2.2657  0.1186  1.1182
#>   0.7885 -5.2852 -5.8395 -3.9467  2.4293
#>  -4.0433  3.1484 -6.0580 -2.8911 -0.0723
#> 
#> (1,7,.,.) = 
#>  -7.1288 -6.2237  4.7839  2.3653  5.8593
#>   2.8874  3.5236 -14.1575 -10.0166 -1.3563
#>  -8.8318 -10.0747  5.1539  2.7909  0.4521
#>   1.1774  6.7935  5.1372  1.5119  0.8252
#>  -1.5564  0.1769  2.6276 -8.1158 -5.4921
#> 
#> (1,8,.,.) = 
#>  -1.0772 -4.9240  1.3226 -5.3752 -4.2907
#>  -1.4143  5.8723 -5.9510 -7.2433 -2.4634
#>   3.5854 -5.4658  0.7022  5.8960  1.1206
#>   6.1263 -4.4769 -12.8672 -2.7106  2.3638
#>   0.2109 -1.6394 -9.4003 -2.6525 -0.6314
#> [ CPUFloatType{1,8,5,5} ]