Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   1.3624 -4.7585  1.7799 -5.7870  2.1500
#>  -3.7862 -4.1257 -7.7875  1.0641 -2.9028
#>   8.3085  5.8666 -0.8243  1.6478 -9.1782
#>  -0.7591  2.8980  1.4486  6.1955 -3.0535
#>  -0.2418 -4.0987 -3.2488 -7.4562 -0.0830
#> 
#> (1,2,.,.) = 
#>  -2.9306  3.9894 -0.6110  6.0279  8.4459
#>   0.5025 -2.9586  5.1465  2.0122  1.8590
#>   1.3737  7.2757  6.7217  2.0162 -0.7830
#>  -5.3296 -0.2344 -1.1449  3.2921  1.9591
#>   5.4167 -1.8539 -4.8965 -6.4474 -6.0883
#> 
#> (1,3,.,.) = 
#>  -5.8486 -3.2859 -4.2095 -0.2003 -1.4495
#>  -5.0800  1.8519  0.0354 -0.8227 -3.4001
#>   0.1207  4.4074  1.7781  4.8339  4.8486
#>  -1.2926 -0.4310  0.0205  0.4052  5.5078
#>   0.9526 -5.4227 -1.8678 -1.9148  2.3831
#> 
#> (1,4,.,.) = 
#>    0.0423   7.0038  -8.7714   6.2114  -2.9504
#>    5.5216  -1.1403   9.6118   5.0192  -0.3835
#>   -3.0947  -4.8219   6.0277  -2.0979   3.8179
#>    1.0135  10.4257  -5.0759  -7.7062  -2.2373
#>    3.1885   1.3115   2.3662   7.5064  -2.7819
#> 
#> (1,5,.,.) = 
#>   -2.2751  -2.1506   4.0901  -8.6614  -0.8002
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]