Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.3662  -0.1117  -6.7326  -1.3479   8.9371
#>   -0.6793   3.8477  -4.1736   5.3499   2.0596
#>   -2.8919   0.1081   3.0909  -4.2485  10.1250
#>    4.5362  -3.4312  11.7753 -11.2349  -8.7622
#>  -12.9409  -3.9887   0.8818 -12.4510  -0.5851
#> 
#> (1,2,.,.) = 
#>  -0.6047 -0.6552  0.4654 -2.5785 -0.1421
#>  -2.9106 -2.0412 -3.6748 -1.8724  8.1270
#>   1.5775 -4.1690 -1.0358  0.6422 -10.8347
#>   3.6614 -10.6373  3.8306 -4.1389  6.4451
#>   6.2960  6.3361 -4.9054 -3.4209 -2.5199
#> 
#> (1,3,.,.) = 
#>    0.3135  -3.0550  -0.9927   1.6550  -0.5809
#>   -0.1667   1.4931  11.1702   2.3493  -0.9825
#>    9.5098   2.9610  -3.2661  -1.9343   2.6128
#>   -7.2736  -2.4628   0.3380  -0.2344  -0.6687
#>    1.0732   3.4124  -1.6610  -0.6129   7.2497
#> 
#> (1,4,.,.) = 
#>   0.4549  1.9346 -1.1532 -3.4431 -0.3919
#>  -3.8139 -2.2305  4.7713 -4.6719  5.8124
#>   2.0451  9.6747 -4.8073 -10.2356 -2.8311
#>   0.8478 -0.5625  5.8462  4.2835 -1.5692
#>   2.8830 -6.4584  4.8694  0.7494 -4.0546
#> 
#> (1,5,.,.) = 
#>   1.0398 -4.9053  1.4188 -0.8690  1.7657
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]