Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -3.2597  0.6219 -1.5623  7.3641  2.3579
#>  -8.9897  7.6611 -4.2115 -0.8596  3.9848
#>  -1.4417 -11.0940  9.7779 -5.9030  2.2080
#>   6.2135 -0.3163  2.7865 -1.1722  4.3569
#>   3.1839 -5.8214  6.6463 -0.1558  0.8794
#> 
#> (1,2,.,.) = 
#>   4.2593 -0.7900  3.6085 -9.7426 -5.9579
#>   1.9354 -4.8053  7.9045 -1.3750  3.6715
#>   5.9058 -1.3142 -6.0911 -6.1655 -3.7185
#>   8.3137  3.9686 -3.5244  1.9630  5.4181
#>   1.5642 -8.3478 -5.7958 -1.3079 -0.9815
#> 
#> (1,3,.,.) = 
#>    2.1539   2.1051  -2.3420  -0.1160   0.0798
#>    8.7688  11.1164  -2.6103  -7.3877  -0.1113
#>    1.5705   0.8419  -0.2830  -0.9450   0.9994
#>    4.8958  11.1447   1.7384  -8.2007   2.4835
#>   -0.8985  -8.1013  11.1407  -0.0346   0.1171
#> 
#> (1,4,.,.) = 
#>   -1.3135   4.4571  -6.5826  10.4959  -2.3448
#>   -4.3151  -0.4081  -1.1347  -2.4696  -1.5766
#>   -1.5213   0.1436  -1.0765   4.1375   2.0214
#>    2.6536  -6.9643   1.9295  -0.2102 -10.1133
#>    1.2179   5.4253  -2.9815   2.0249  -1.7259
#> 
#> (1,5,.,.) = 
#>  -3.3541  3.9970  3.2096  0.7267  0.3678
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]