Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    1.0067   2.5188 -11.3647  -0.8438   6.9791
#>   -9.2852   7.8023  -7.6124   1.2237   1.4602
#>    4.8710   8.9074  12.4002 -14.3726   4.0842
#>   -3.0880  -6.9015 -10.5938   3.2119  -3.6056
#>    0.3375 -13.6508  -5.1306   4.7523  -2.2722
#> 
#> (1,2,.,.) = 
#>    2.0945  11.0661  -8.9451   1.4401   1.0562
#>   -2.7480  -0.7367   6.2048  -3.3074   2.9803
#>    2.7331   7.3852  -2.7329   7.2912  -3.4701
#>    2.1200   2.3181   5.0959   3.4498  -4.8826
#>    4.0965   3.4483  -0.3662 -10.1702  -1.0121
#> 
#> (1,3,.,.) = 
#>   -5.1321   3.1087   3.1419  -0.4268   2.6835
#>   -7.7446   0.9038  11.3450   0.1397   2.6630
#>   -4.6575   7.5739  -5.6445   1.1188  -2.0086
#>    4.9237   1.8554  -4.3465  -8.5235   2.9581
#>    5.5354   9.6856  -3.5698  -8.0198  -1.2322
#> 
#> (1,4,.,.) = 
#>   3.0569  0.0556 -3.8294 -3.9816  1.3477
#>  -9.6004 -1.3463  3.1870  2.5492 -2.1309
#>   6.1954 -4.6334  2.9131  1.9174  1.7573
#>  -4.8279  3.5893  0.2824  1.1824 -5.4654
#>   1.5668 -2.2086 -3.2305 -3.6714  3.5496
#> 
#> (1,5,.,.) = 
#>   12.2200   1.1128  -9.8876   4.3807  -0.7622
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]