Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   0.7151  5.6989  4.4157  2.7929 -1.3940
#>   5.2416 -11.8660 -5.7111  3.7804 -3.6648
#>   6.7004  4.2566  1.2996 -3.6517 -2.0848
#>  -3.2391  1.8586  6.5147 -3.8470 -1.0240
#>   4.8523  2.2899  4.9731  2.6813  5.3291
#> 
#> (1,2,.,.) = 
#>   1.3253  3.0580  2.6331 -9.2044  0.6648
#>   3.3940  6.4863  0.1899  7.7943  6.4709
#>  -4.8129 -5.8083  2.2095 -0.5284 -4.9605
#>  -0.9617  0.4091 -5.7827 -1.4512 -5.9941
#>   1.0555  0.2167 -5.4010  5.1052 -1.4834
#> 
#> (1,3,.,.) = 
#>    0.6698  -0.1128  -0.2921 -12.3713   3.7687
#>    0.8852  -0.3766  -1.6016   3.9388  15.1976
#>    5.6880  -2.5243  -8.0063  -0.8307  -1.8462
#>   -1.6116  15.0529   4.7998   0.0960   2.2226
#>   -3.5625  -7.0019  -6.2608   7.7775   0.6235
#> 
#> (1,4,.,.) = 
#>    3.8237   5.0783  -9.1538   9.0535   9.7278
#>    1.9179  -6.1586  11.4054   3.2508  -4.7829
#>    7.3312   2.4353  -1.5627  -4.4059  -6.3160
#>   -0.9648   5.4968   5.1252  13.4040   2.4816
#>   -0.5396   1.2242  -3.2759  -2.2843   1.5039
#> 
#> (1,5,.,.) = 
#>  -2.7906 -9.0891 -1.9540  9.1021  3.8549
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]