Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    5.4659   2.8990   5.3984   7.0360  -2.0090
#>   -5.9133  -9.8886  -5.9084  -5.1765  -1.7190
#>    2.7944  -4.0962  -1.5758  -7.5222  -3.4789
#>    4.4229  -9.8143  -1.1076   8.8076  -3.9584
#>   11.9805   6.1494   2.6441   0.9381   0.2609
#> 
#> (1,2,.,.) = 
#>   2.9906  6.1203  3.5234 -4.1654  2.7427
#>   8.6287 -1.1923 -0.9438  5.4518 -5.8046
#>   2.4781  0.5789  2.8303  6.8565 -12.2146
#>  -2.6909 -3.4307  1.0314 -13.2807 -1.5225
#>   4.6936 -7.2498  6.0408 -2.4307 -3.0278
#> 
#> (1,3,.,.) = 
#>   0.4152 -0.9448 -7.4997 -10.3879 -0.1850
#>   2.2398 -4.3109  3.1288  3.4429  5.6086
#>   4.6241 -12.4299  2.3733  1.2171  0.4450
#>   1.3748 -7.6937 -2.2415  9.8533  4.6029
#>  -5.0836 -4.9363 -0.3118  2.3015  2.7628
#> 
#> (1,4,.,.) = 
#>    4.4001   6.4772  -1.5499   0.4220   2.8249
#>    4.1705  11.8581   3.8433   4.4023   0.8681
#>    1.7160  -3.8402  -3.9065 -13.4539  -4.4388
#>    2.8702  -6.2747  -3.0816  -4.6862  -5.1075
#>   -0.0767   6.8535  -3.0770   1.0034   1.2959
#> 
#> (1,5,.,.) = 
#>  -0.9179  7.4870 -1.5710 -0.0047 -7.1386
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]