Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   0.0028  4.7855 -1.0971  0.5028 -0.4350
#>  -3.9280  3.9926 -1.6503 -0.6803 -0.5539
#>  -0.4285 -3.4354  0.8228  3.7908 -7.0264
#>  -4.3371 -2.9409  9.7713  9.1871  1.3260
#>  -1.7715  2.8257 -0.4352  0.9805 -1.6010
#> 
#> (1,2,.,.) = 
#>   6.8315 -1.1977  3.5481 -5.8067  1.7606
#>  -2.4894  6.5683 -2.7309 -0.2708 -0.2657
#>  -3.1254  1.2116  3.0541  3.4977  6.4309
#>  -1.8635  5.2934  2.3297  3.9655 -7.2980
#>   1.6660  0.4504  3.9031 -3.2873 -1.4390
#> 
#> (1,3,.,.) = 
#>   -4.3112  -3.0059  -0.9606   1.6890   2.8577
#>   10.9183   7.4454  -0.0498  -6.4501   2.0293
#>    0.2141 -11.7206   2.6579   6.3070  -4.3483
#>   -0.1014  -0.9514  -3.3162  -0.8143  -4.3073
#>    2.4183   0.4169  -3.7370   1.2811   6.5875
#> 
#> (1,4,.,.) = 
#>  -6.1327 -6.9306  3.3813 -3.9926  4.1657
#>   2.3335  0.8546  9.2176 -7.2219  2.4416
#>  -5.1488 -8.4397 -6.9719 -3.5853 -6.8458
#>   1.1478 -1.7885  6.0930 -0.1825  0.4976
#>  -1.4591 -1.1058 -1.3027 -3.0503  4.4008
#> 
#> (1,5,.,.) = 
#>    7.6443  -8.7673 -10.3452   3.2486  -0.2407
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]