Conv2d

Conv2d

Arguments

input
NA input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
NA filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
NA optional bias tensor of shape \((\mbox{out\_channels})\). Default: None

stride
NA the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
NA implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
NA the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
NA split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=None, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See ~torch.nn.Conv2d for details and output shape.

.. include:: cudnn_deterministic.rst

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -4.9664 -1.8468 -1.1927 -5.6745  3.0247
#>   4.3059  3.3988 -0.5044  9.9966 -2.0411
#>   6.3643  4.8541  3.8433 -3.5750  8.9129
#>  -0.0392 -7.1811  7.0310  4.0186 -0.8507
#>  -9.9586  1.9385 -2.9976  8.9305 -3.3952
#> 
#> (1,2,.,.) = 
#>   0.9641  0.3437  4.3658 -0.0712 -4.1856
#>   4.4279  3.0379 -4.7309  6.2338  8.3105
#>   2.0309 -1.0563  9.1169  0.0176 -3.1387
#>  -8.7372 -3.4841 -11.6829  9.1988 -1.9338
#>   1.0892 -1.8982 -4.3750 -0.6017  4.9374
#> 
#> (1,3,.,.) = 
#>    3.3299  -3.5973   0.3394   8.2893   5.7314
#>    2.0527  -0.1722  11.8906   2.3552  -0.0871
#>   -5.9784  -3.4574  -0.6590  -1.5255   4.5437
#>   -1.9965  -1.8227  -1.7939   0.1806  -0.1500
#>   -1.9760   6.1899   0.6041  -1.6072  -0.3464
#> 
#> (1,4,.,.) = 
#>   5.5668 -6.1740  8.9316  4.3312  0.9933
#>   3.1101 -1.7972  9.8388 -5.2957  6.6138
#>   2.1642 -9.1746  6.3721 -8.8527 -1.5292
#>  -10.7355  0.1007  2.2665 -3.7808  3.1730
#>   1.4664 -1.5943  4.6338 -3.0332 -3.0062
#> 
#> (1,5,.,.) = 
#>  -0.2230 -5.5796 -1.0343  0.7457 -0.3506
#>  -1.3786 -0.5930 -2.2193 -3.3161 -4.1868
#>   3.2025  5.8686 -7.0942 -2.1347 -1.8191
#>  -3.9774  7.4350 -0.6601  5.2576 -6.5332
#>  -2.3488  2.2309 -1.2750  4.4132 -3.0113
#> 
#> (1,6,.,.) = 
#>    6.0492  -1.1223   7.6921   5.0029   3.0605
#>  -10.6686   2.1366  -4.1249  -6.0583   5.9899
#>   -6.6301 -15.9595  -0.6232  11.0436   6.6455
#>    3.8833  -1.8381  -0.6667   4.0604  13.0146
#>   -1.5805  -5.9410   1.4505   0.0443  -6.2604
#> 
#> (1,7,.,.) = 
#>   -4.6263  -1.8233   0.9582   4.8793   3.4146
#>    1.0303  11.2821   3.2465   4.7268   1.7562
#>   -6.3917   5.2913 -17.5528  -2.4286  -6.8908
#>   -5.0196   4.0957  -4.2696  11.5655  -4.1123
#>    6.6294  -4.2749   4.3220  -6.5068   4.3931
#> 
#> (1,8,.,.) = 
#>   -1.9190   1.2976   8.3696   5.0672   5.5601
#>    3.1500   9.5757  -1.1553  -5.3542  -8.0591
#>  -10.6190 -10.4182  -8.6754   2.8377   1.6882
#>    9.3264   4.7451  -1.0924   3.8493   0.5546
#>   11.5317   3.1660  -4.2486  -7.5473   5.0758
#> [ CPUFloatType{1,8,5,5} ]