Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -1.2906  2.9120 -10.1347 -8.4063 -2.7954
#>  -8.4893 -0.6753  4.2519  5.4861  0.5061
#>   1.2696  3.6413 -0.4040  6.2206  0.8949
#>   5.6457 -2.5196 -1.7400  8.3430 -5.1519
#>  -0.1973  6.0974 -6.8863  6.3241 -4.3806
#> 
#> (1,2,.,.) = 
#>  -3.6324  0.3120 -2.5309 -1.3126 -0.2244
#>   7.9885  5.0540  2.1115 -10.2176  0.0942
#>   2.7272 -3.6948  4.2059 -8.1242  1.8549
#>  -2.2596  9.6812 -1.4564  2.2219  0.4378
#>  -5.8727 -1.9664 -1.4101 -7.6570  2.8877
#> 
#> (1,3,.,.) = 
#>   2.1833  3.2364  0.2910 -0.1084 -2.5001
#>   0.8589 -5.8714  4.3053  4.9777  5.3098
#>   8.8078 -1.0887  3.8973  5.5964 -1.0706
#>   1.7158  0.6410  0.3895  5.2919 -0.9719
#>   3.1001 -5.0140 -3.4551 -1.3446 -5.8430
#> 
#> (1,4,.,.) = 
#>   -3.1360  -2.7810  -4.1115  -1.7909   4.0383
#>    0.4955   5.7118  14.2705   7.3208   2.0902
#>    4.5326  -0.9656   0.8890   0.9597   1.0325
#>   -1.0812   3.1205   5.0999   1.4562  -1.3270
#>    0.4626  -5.6533   4.0904  -0.9684  -3.0868
#> 
#> (1,5,.,.) = 
#>  -1.1100 -2.5651 -4.2485  2.2354  0.8958
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]