Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    3.1625   5.3163  -5.5938   0.3845  -6.3885
#>    6.5694  -5.7123   1.7777  -0.6199  -5.4881
#>    5.0001  -4.4543  -7.8152  -7.0796  10.9347
#>   -2.4959   7.0330   0.8140  -5.9283  -1.7393
#>   -0.9232  -2.3069   4.2536   3.3093  -1.6377
#> 
#> (1,2,.,.) = 
#>   4.9945  3.6298  2.4018  0.5479 -3.8883
#>  -5.3224  1.5096  9.4223 -1.8798 -4.2023
#>  -5.2506  5.5988 -6.1448 -1.8679  5.8568
#>  -0.6203  2.0162 -8.2463  9.5398  1.6239
#>  -11.8652  4.6003  6.5386  1.6412 -5.3118
#> 
#> (1,3,.,.) = 
#>   -5.0496  -5.4262  -3.0320   3.3295  -3.0136
#>   -2.9717  -0.8810   1.9084  -9.4960  -7.8053
#>   -4.1063   3.6525  -2.0170  -1.6359   6.7578
#>   -0.1688  13.0495 -11.2772   1.8443  12.4935
#>   -2.9743   9.4682  -1.0594   2.8032   0.0483
#> 
#> (1,4,.,.) = 
#>  -3.7862  2.3248 -1.1813 -0.6707  0.1328
#>   5.0005 -3.3032 -5.1983  2.2442  5.6947
#>   6.4931 -10.4821 -3.4264 -4.2132 -2.6582
#>  -2.1033  5.1792  4.9643 -5.6829 -0.4048
#>   7.3828 -14.2976  5.5978  3.6778 -5.9191
#> 
#> (1,5,.,.) = 
#>   -0.4276  -1.5092  -4.7544  -5.7866   2.8864
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]