Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   0.2425  3.5073 -5.1447 -6.6398  1.5635
#>   0.0951 -2.9432 -6.4521  4.5018 -5.6029
#>   2.8747 -4.5687 -10.1055  3.4469 -0.7040
#>   0.7106 -1.8919  5.1157 -3.2080  3.0289
#>   2.5211 -4.1812  7.7407 -3.4750  1.3203
#> 
#> (1,2,.,.) = 
#>  -0.2888  0.2048  2.1611 -0.6698  4.5137
#>  -6.1176  3.7332 -3.9477  2.0802  1.6047
#>   0.2170  4.6105  0.3302  2.2807  0.0048
#>  -0.8543 -0.1787 -6.4197 -7.9117 -2.7084
#>  -2.7177  1.1437 -2.2770  1.0009  0.9719
#> 
#> (1,3,.,.) = 
#>  -3.0723  2.8163  2.5302 -5.8506  1.6172
#>  -5.9853 -5.1816  4.5535 -1.4597 -1.6599
#>  -1.7309  3.5511 -1.6846  5.3960 -5.3175
#>  -3.1692  1.5223 -1.1949  6.0913 -3.6426
#>   1.4080 -4.6918  5.1207  2.3421 -0.1075
#> 
#> (1,4,.,.) = 
#>  -4.4284  2.2310  0.7276  4.8435  3.0760
#>  -2.9497  5.9539  0.2245  7.9617  6.1158
#>  -5.6720  2.6683 -6.3417  4.9378 -3.8963
#>  -1.7421 -0.9693 -2.3879 -2.0791  2.8129
#>   4.2450 -0.2581  4.5776 -2.2884 -0.5702
#> 
#> (1,5,.,.) = 
#>   3.7464  3.5371 -4.2673  5.4492  4.2325
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]