Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.0525  -0.3950   0.3368  -2.6956  10.9064
#>   -2.7875  -9.3135  -1.3835   2.3648   9.0541
#>    3.1540   2.3282  10.5431   7.2890   2.0269
#>   -3.1850  -3.6807   1.0499  -8.4626  -9.3757
#>    0.3697   1.0526  -4.1858   4.9227   1.6459
#> 
#> (1,2,.,.) = 
#>    5.4778  -3.9899   1.2618  -5.1641  -3.7144
#>    2.9995  -0.6043   6.2648  -9.4356  -7.0800
#>    6.4513  -6.0746   4.0601  -4.6841   3.2939
#>   11.2861   7.3077  -1.7582   2.9226   2.6630
#>    8.6236  -3.7549   1.3866   4.2060  -5.7672
#> 
#> (1,3,.,.) = 
#>    0.7788  -9.1280   6.7275   2.9052   3.5484
#>   -0.8069  -0.8149   0.7337  -5.6034  -2.3626
#>   -0.4832  -8.7056  -5.3246  -6.0841  -3.2847
#>   -5.6473   0.5160   4.8168   3.3827  -3.1157
#>   10.5564  -1.9186   4.0956  -6.4395  -3.9016
#> 
#> (1,4,.,.) = 
#>    7.2975  -2.7724  10.1590  -1.7053   1.2597
#>    1.6083  -2.5077   6.1173   3.0141   1.4391
#>    6.5150   1.2012   8.3187  15.5836  -4.9788
#>   -1.6480   2.9907  -6.1918  -2.1582  -2.8335
#>   -4.9956  -2.5404   4.7907  -1.5278   8.4529
#> 
#> (1,5,.,.) = 
#>    1.0819   9.0654   0.9083 -11.3267   2.2492
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]