Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    0.4592  -6.7283  -4.9136  -5.5847   1.6627
#>   -2.4434   5.6380   6.0609 -11.6053 -13.3588
#>    2.6099  -1.9502  -4.3057  -3.7867  -0.1148
#>   10.0197 -10.8825  -4.6743  16.0190  -0.7299
#>    1.8553  -7.7312   7.8126   2.2743  -7.8414
#> 
#> (1,2,.,.) = 
#>   -4.2469  -8.1312  -6.7837  10.7211  -1.0717
#>   -0.9223  -4.6656  -2.0406   9.7520   2.3030
#>   -1.0812  12.1711  -3.9165  -4.2127   3.0310
#>    3.8167  -2.0450   2.7122   6.5828   3.6392
#>   -3.2698   6.3987   3.8213  -7.3366  12.2806
#> 
#> (1,3,.,.) = 
#>  -2.4796 -3.6590 -4.8961 -0.2459  2.6119
#>   2.0117  6.4899 -0.8231 -6.3675  4.1601
#>   1.1558 -2.6383 -0.3941 -1.1117 -1.1466
#>   0.4906 -4.3037  7.2306 -2.3458  4.8067
#>   1.8806 -4.2582  1.8817  1.4462 -0.2938
#> 
#> (1,4,.,.) = 
#>  -4.2560  3.5808  0.9534 -3.9417  4.2230
#>   2.1786  3.8573 -1.4491 -8.0715  2.0657
#>  -2.5176 -4.0630  0.8485 -4.1503  2.6798
#>   2.4669 -6.2978 -1.8435 -2.6458 -7.8071
#>  -2.5664 -1.7640  6.1281  0.7712 -6.2965
#> 
#> (1,5,.,.) = 
#>   -7.3272  -5.0129 -12.9358  -4.6108   3.4105
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]