Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   -7.2444   5.7316  -3.4137   0.4194  -0.8337
#>    7.4948   6.1236   1.3580   6.5343   5.7662
#>   -3.4971   1.6228  -3.4571  -2.1898   1.9390
#>   -1.6602  10.5275   1.8648  -0.0782  -0.4030
#>    0.4240   2.1998   0.5763   3.0213   4.9353
#> 
#> (1,2,.,.) = 
#>   1.7498 -11.3319  5.2025  2.5297 -12.3158
#>  -0.0866  4.7865 -5.8977 -1.3142 -1.9935
#>  -9.2358 -8.0456  0.6234 -4.8209  6.9196
#>   4.4150 -5.8079  2.4874 -1.2597 -4.6401
#>  -1.2831 -0.8067 -4.5058 -1.6990  1.4104
#> 
#> (1,3,.,.) = 
#>   -2.4752   6.8377  -2.0029  -5.7994   4.4840
#>    5.7412   5.2225  12.8323   8.9389   0.1804
#>    0.1393  -4.2034  -2.4708  -5.7871  -2.5713
#>   -1.4268   6.2611   5.5541  -4.4341  -0.4154
#>   -0.2290  -1.9387   3.8245   5.1016   2.7239
#> 
#> (1,4,.,.) = 
#>    9.6455   3.1221   4.2455   4.1184  -7.1496
#>    1.9892  -2.1039  -4.4529  14.3328   5.9729
#>   -0.2884  -7.2423   7.1994   1.3553  -7.1051
#>    7.8007  -6.2863   1.7525  -5.7006  -3.4389
#>   -7.2390  -3.3792  -0.2922   4.7075   2.8058
#> 
#> (1,5,.,.) = 
#>    4.1557  -2.8762  -3.8542  10.9322  -6.1805
#>   -6.7864   3.5341   0.5644   5.5398   6.6681
#>   -4.9332   6.2398  -1.3390 -11.6520   7.2444
#>    3.3821  -3.8478  -2.6133   3.3345  -6.6994
#>    2.0202   7.1868  -0.7253   7.4965  -2.5467
#> 
#> (1,6,.,.) = 
#>   -1.2660  -7.0724  -3.7131   1.7050  -2.9775
#>   -0.3792  10.4570   2.1986   4.7298  -5.7242
#>    3.2162   1.2338   9.7716  -1.9126   8.7052
#>   -5.1104  -6.1430   2.7566   5.8639   6.2763
#>   -0.1262   7.2134  -6.4083   0.4634  -5.5659
#> 
#> (1,7,.,.) = 
#>  -7.5180  0.1238 -8.1245  0.5265 -0.1994
#>   1.9569 -1.7652 -2.0653 -7.2121  5.4941
#>  -6.5407 -0.2806 -9.5595 -5.5248  4.8217
#>   5.3590  1.8602 -1.1588 -8.4894 -5.4160
#>  -3.6694 -2.5897 -1.8725  0.2242 -0.7954
#> 
#> (1,8,.,.) = 
#>  -4.4078  3.3362 -0.8670 -10.3200  0.2147
#>  -2.5683 -15.8128 -0.8906 -9.6601 -9.7293
#>  -0.9660 -0.5193  1.8374 -0.1317 -5.8345
#>   0.5586  0.0739  1.3577  0.5976 -4.1768
#>   1.3778 -6.7510  0.1554 -5.0435  3.1046
#> [ CPUFloatType{1,8,5,5} ]