Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    3.6691   0.3120 -10.6453  -5.8935  10.2013
#>   -1.2648  11.3883   1.7619  -9.0949  -3.7262
#>    5.9419  -5.6127   3.7024   6.8109   2.0081
#>   -2.1890   5.1721  11.7591   0.1656   2.7924
#>   -1.3161  -1.6539  -3.5556   5.1099  -1.4889
#> 
#> (1,2,.,.) = 
#>    3.9400  -0.8646  -0.3011  -7.2579   8.1667
#>    1.3377  -0.1094   3.9428  -3.2528  -2.9536
#>    1.3926   3.2794  -2.9041  13.7937  -1.9004
#>    0.6780  -3.3739   2.3584   0.1481   3.7224
#>    6.2518  -5.9718   4.6218   4.2902   1.7215
#> 
#> (1,3,.,.) = 
#>    1.1520   9.2964   0.0468   6.5545  -6.6203
#>   -2.3572  11.4622  -5.8189   7.3054  -2.2128
#>   -7.7213  -2.7678   5.3890   5.5175  -7.0218
#>    3.0807   5.2116  -5.2247  -3.9853   2.6967
#>   -6.5988  16.6708   2.5344  -2.2722   6.7204
#> 
#> (1,4,.,.) = 
#>  -1.9175  4.7267 -4.3389 -5.7980  3.9730
#>   3.8769  5.0733 -6.0614 -4.9733 -2.2278
#>  -3.7304  0.7068  0.3114  0.9686  2.7334
#>  -3.9584 -6.5941 -1.3695 -0.0896 -2.3557
#>  -1.6432  5.5549  3.1018 -5.0597  7.5027
#> 
#> (1,5,.,.) = 
#>   6.0208  5.9535 -9.1436 -6.8543  8.4664
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]