Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   0.6015 -1.2146 -3.1188 -1.2386  2.5811
#>  -0.6596 -2.0392 -4.1941  3.3347 -4.8658
#>  -2.9597  2.8568  3.4543 -7.7694 -1.9058
#>   0.5870  1.6258  1.7425 -3.6310 -1.9332
#>  -2.4965  4.4640  2.9169 -5.9571  4.5557
#> 
#> (1,2,.,.) = 
#>    9.4832  -5.9925  -2.4275  -1.6187   3.5308
#>    2.0242  10.3078   5.8800  -2.9517  -4.9437
#>   -3.8268  -3.7981   7.3904  -6.5159  -0.8139
#>   -1.4339  -1.7707   4.1897   2.5826   2.7684
#>    0.5375   3.2882  -1.9920  -2.0343  -7.4479
#> 
#> (1,3,.,.) = 
#>   11.0629   4.2166   3.1659   1.3936  -0.9338
#>   -2.6211   6.4280  10.8596   6.0221  -0.2024
#>   16.1546  -0.4075  -0.0221   6.5756   1.4001
#>   -4.6268   7.9868  -6.7772   2.7896   7.7107
#>    2.7588  -0.9301  -5.0942  -5.0847   2.2749
#> 
#> (1,4,.,.) = 
#>  -5.0842 -7.3362 -11.9168  4.5385 -12.1532
#>  -0.4909  0.6206  5.8431 -6.6750  0.8982
#>  -3.3432 -0.4807 -1.8916 -0.3577 -0.0967
#>  -2.3799  0.2154  2.4420  4.5807  6.8865
#>  -2.0708  4.5270 -4.8910 -10.2053  0.3764
#> 
#> (1,5,.,.) = 
#>    4.1612  10.5958   1.2503   1.1226   0.5789
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]