Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   4.6067 -5.1447 -2.2687 -5.5652 -5.8458
#>   9.0570  5.9004  8.2770 -10.5873 -0.4662
#>  -2.6685  8.7924  2.2810  3.9831  2.0377
#>  -0.6493  5.7815  1.3456  3.8772 -2.8015
#>  -0.8665 -7.4957 -3.0097 -4.1334 -3.3825
#> 
#> (1,2,.,.) = 
#>    3.4063  -5.4509  12.6341  -7.7464  -1.7827
#>   -0.8976   3.3401  -7.4160  -4.6649  -5.3346
#>    5.6834  -1.6107  16.6736   8.6876  -2.6554
#>    0.9408  -2.0604  -2.3721   7.0779  -2.2152
#>    2.8424  -2.0587   7.2388   1.1752   2.3546
#> 
#> (1,3,.,.) = 
#>   -7.5345  15.2359 -13.6099  12.2935   0.5121
#>    1.7205   0.1960 -12.3255  20.0697  -5.3078
#>    2.0395  -2.7668  -8.4097   3.2234 -11.2627
#>    5.4729  -8.7635   1.6928   2.3429   2.9791
#>   -9.7048   1.7376  -6.3314   1.0566  -7.3306
#> 
#> (1,4,.,.) = 
#>  -4.0896 -4.7108  0.7781  5.4911  0.5307
#>  -0.2002  1.7459  9.8369 -0.4697  0.2562
#>   3.0494 -12.2151  2.5999  7.2787 -4.9914
#>  -4.0951 -7.6884  4.7729  4.8562 -3.8038
#>  -0.5653  1.5257 -1.4235  5.3537  2.5916
#> 
#> (1,5,.,.) = 
#>   -4.5981  13.3623   2.0465   5.6323  -0.5039
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]