Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -2.2563  4.3531 -1.9905 -3.2304  1.3243
#>   2.2112  8.3018 -1.8664 -6.2126  1.1917
#>   4.6528  9.0252 -3.1500  3.8881  7.4830
#>   6.1715  3.9133  0.7480 -1.4733 -4.6361
#>  -1.1611 -3.6615  0.1511 -3.2846 -2.8228
#> 
#> (1,2,.,.) = 
#>   -2.5341   1.6122  13.1917  -0.5329  -3.1153
#>    2.4824   1.9005  10.4470  -4.8168  -1.9034
#>   -5.8270  -0.9094   2.4674   3.4901  -1.4828
#>   -2.9038  -4.5101   6.6154   7.4477  12.1339
#>   -1.8350  10.1810  -8.1237   2.4943  -4.9625
#> 
#> (1,3,.,.) = 
#>  -3.6606 -0.9685  7.3020  4.7810 -2.8099
#>   6.0437 -6.8014 -0.0931  4.1982 -6.0696
#>   6.2086  3.5209 -7.3890  7.6013 -8.5866
#>   7.9358 -6.2882  0.5888  5.8255 -3.8177
#>  -2.6680  0.8882  0.5712  5.5644  2.8463
#> 
#> (1,4,.,.) = 
#>  -0.0101  5.0494  1.5536 -5.7925 -8.5564
#>  -0.7877  0.7436 -3.0538 -9.1286  9.9943
#>   3.6592  2.0427 -2.3913 -7.8006  1.5784
#>  -1.1165  2.4095 -1.5952 -10.5022 -9.4910
#>  -0.5394  8.6777 -6.3766  4.1196 -10.1166
#> 
#> (1,5,.,.) = 
#>  -4.8988 -1.8468 -0.1367 -0.9379 -6.4342
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]