Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -1.5009 -3.8619  1.4198  7.8090  1.1869
#>  -4.4909 -0.7102  2.1434 -4.4441  6.3492
#>  -1.5354 -4.2128 -2.1573  0.9706 -4.9769
#>  -2.8455  3.1226  4.8831 -3.3034  2.3631
#>  -1.1566  1.4397 -0.9804  6.5530 -4.3531
#> 
#> (1,2,.,.) = 
#>  -2.6077  4.2568 -2.6337 -5.1881  0.9652
#>   0.8390 -1.1920  3.4392 -3.5270  3.0800
#>   2.7486  6.0991 -0.5892 -5.9834  2.0411
#>  -5.2374 -0.2376  4.4895 -5.1448  0.1984
#>  -0.7269  0.4938  3.5713  2.6290 -1.1511
#> 
#> (1,3,.,.) = 
#>   -3.3375  -1.9187   1.2180   2.6142  -0.1526
#>    1.5150   9.0596  11.2671   2.6194   2.5530
#>   -1.7726   4.2292   4.5631  -5.4741   3.3617
#>   -4.0220  -7.4068   2.9742  -4.7044   4.2746
#>   -5.8090   5.4452   1.1106  -2.5566  -1.1970
#> 
#> (1,4,.,.) = 
#>  -3.4854 -13.7033  0.5490  4.7655  5.2162
#>  -1.4458  4.1741  6.0986  6.8531  0.4519
#>  -8.5474  0.9887  0.6212 -6.3003 -1.8396
#>   5.5567 -2.0197  1.6614  9.0434 -1.3302
#>  -3.1776  3.3387 -0.0387  3.0613 -1.0215
#> 
#> (1,5,.,.) = 
#>    2.4827  -2.5805   0.3477  -3.7075   1.3948
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]