Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   6.5474  1.2523 -4.3449  3.8304 -6.6515
#>  -4.5053 -10.8437 -2.0629  0.1279 -0.7558
#>  -1.1004 -2.1481  4.4085  1.9975 -2.7054
#>   2.5377  6.1088  1.1038  5.4126  2.4067
#>   1.4163 -0.6436  2.6925  6.2138  5.8480
#> 
#> (1,2,.,.) = 
#>   4.3473 -11.7102 -3.3430 -2.7155 -2.1801
#>  -1.9665 -0.9455  9.3554 -5.9011  8.5120
#>  -5.2958  7.2555 -9.4118  6.5831  0.1966
#>  -7.9475  7.6014  4.6462 -9.0019  2.1333
#>   2.0538 -6.3202  8.5200 -5.2738 -7.5918
#> 
#> (1,3,.,.) = 
#>  -8.2913 -2.1968  5.7042 -0.8133  0.8151
#>   1.3434 -0.0744  2.6706  7.9648  2.3526
#>   6.3036 -2.1109  2.8829 -3.1263  1.3497
#>  -2.7845  2.5146 -6.4455  6.1990 -6.2240
#>  -1.1417 -1.5141  0.0120 -2.8793  3.7893
#> 
#> (1,4,.,.) = 
#>    0.2088  -2.6840 -11.4922  -0.5644   1.5903
#>    7.5801  -2.2262  11.9552  -1.3349  -0.3951
#>   11.3165   8.1192   1.3777  -0.3609  -0.2850
#>   -7.6444  -2.1123 -14.1002  11.4281   8.8798
#>   -7.7459  -4.4947   4.8958   5.5733   5.2437
#> 
#> (1,5,.,.) = 
#>   -5.0064  -3.5850   4.5469  -7.7242   3.9667
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]