Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -3.1119 -1.5098  1.3188  1.3458  1.4775
#>   2.8681 -7.7562 -0.9657  1.9437 -1.8327
#>  -9.0116 -2.0892 -8.1699  1.8463 -2.5261
#>   0.5935 -3.6998  2.4474  2.6212  1.2536
#>  -2.7278 -5.6281 -3.9814 -5.1791  0.1246
#> 
#> (1,2,.,.) = 
#>   2.1052  1.9264  2.1547  2.9431 -2.1518
#>  -3.7710  4.4226  4.0253 -2.6713  1.5648
#>  -2.6851 -3.3330 -3.8507 -1.1707  1.3124
#>  -2.2679 -3.7123 -4.8384  7.8061 -2.9655
#>  -0.5516 -3.9898 -2.0610  1.2049 -5.6531
#> 
#> (1,3,.,.) = 
#>    7.6987  -7.8596   8.0271  -2.5053  -0.4767
#>   -2.6845  -7.1725   8.8504 -11.1061   0.4311
#>    2.9818   8.7313   1.1198  12.1307  -0.2330
#>   -0.1244  -8.7161   0.6273   2.0650   3.2182
#>    1.9803  -0.0803   8.0005  -1.1814  -1.5267
#> 
#> (1,4,.,.) = 
#>   -2.9960  -0.7365 -11.4556   2.1872  -3.4944
#>    8.2163   0.8217  -1.7831  -2.8608  -3.4086
#>    6.9778   1.5144  -0.7539   2.2106  -3.5088
#>    3.3643  12.5373   0.0481  -9.6190  -2.5819
#>   -1.5524  -2.1215  -2.8586   6.8023   4.4643
#> 
#> (1,5,.,.) = 
#>  -10.0212   4.7790   3.6214  -0.5627   0.7928
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]