Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    1.7787  -6.4899   6.5776  -1.4198  -2.3396
#>   -2.9332  -0.7816  -3.6806   2.3501  -7.9953
#>    0.5113  17.6628   1.9692   6.7594  -5.5143
#>   -3.3263  -2.8457  -2.2659   6.1619   0.1797
#>    3.5016   2.0881  -1.6267  -3.9164   3.5264
#> 
#> (1,2,.,.) = 
#>  -4.7390  2.6057 -1.2816 -3.9206 -7.4658
#>  -1.3492  4.8769 -7.0508 -4.1375  0.3861
#>   2.9366  4.1508 -9.3241  5.9644 -4.7627
#>  -0.6816 -2.9139 -3.6415 -4.3299  3.1909
#>  -9.0582 -9.7718 -0.7450 -0.5283  7.3300
#> 
#> (1,3,.,.) = 
#>   0.2405 -3.3067  3.4465 -5.1104 -1.8751
#>   2.2134  5.4905  6.9432 -8.1671  1.5949
#>  -4.3940  6.8259  0.4199 -2.8125  5.4827
#>  -2.1366 -1.1483 -1.1335 -5.8721 -1.4991
#>   1.4581 -8.2515 -3.5140  2.4545  4.9958
#> 
#> (1,4,.,.) = 
#>    6.0945   7.2564  -9.1684   2.7660   1.0189
#>   11.6207  -9.1009  -7.2545  -0.3072   4.2815
#>    1.0309  -6.0071   0.7873  10.3126  -4.6239
#>   -7.3639  -6.7071  -1.1241   2.9127  -7.1878
#>    3.1181   5.1685   5.4143  -8.3894  -0.8710
#> 
#> (1,5,.,.) = 
#>   -5.7580   5.8824  -4.2918   9.5348   1.7807
#>    5.3352 -14.2144  13.6167  -4.1318  11.1063
#>    4.4839  -1.3516   3.4877   3.2574  -8.4403
#>   -6.2358   6.8403  11.7978   2.4195   2.8107
#>    6.9124   2.6935   6.4752   0.8845   4.5259
#> 
#> (1,6,.,.) = 
#>   6.1574 -6.6648 -0.1032 -7.1657  1.0046
#>   3.5231 -0.6414 -1.8246 -1.9976  5.0977
#>  -1.4226 -3.8935  3.8253 -0.5414  4.0582
#>  -2.2777  0.8748 -1.6178  3.2799 -0.1681
#>  -3.1439 -0.5489  1.9070  1.4624 -7.3411
#> 
#> (1,7,.,.) = 
#>   0.5635  2.4170 -0.2178  4.9947 -1.8311
#>   4.6529 -4.8008  4.5786 -11.2098  7.7544
#>   0.8445 -8.8711  8.6925  4.7697 -3.6078
#>   1.4985  6.3979 -4.2185  2.8709 -7.4708
#>  -0.1878 -2.1565  4.0968  2.6758 -2.8600
#> 
#> (1,8,.,.) = 
#>    4.5860  -6.2194   2.7488   4.5865   0.6195
#>   -0.9079   2.8323  -2.2463   3.6367  -0.0743
#>    3.0497  -1.2445   7.7099  -0.5153  -0.2538
#>    0.5601  -4.6063  14.5756  -5.3472   0.0898
#>    2.4165   2.5990   4.2575  -1.4401  -2.5121
#> [ CPUFloatType{1,8,5,5} ]