Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -2.2396 -9.2057 -4.0538 -0.4565  0.2009
#>  -5.5384  2.1571 -2.6272  2.7460 -8.1978
#>  -2.0855  2.4472  3.9472  8.0338  4.0936
#>  -1.1878 -4.8492 -0.2720  6.4239  5.7832
#>  -1.4442  1.3141 -5.1097  0.0983 -0.3734
#> 
#> (1,2,.,.) = 
#>  -2.3233 -0.6653  0.7724  0.4012  3.0529
#>   1.4904  1.1899 -0.3882 -9.6357 -2.7067
#>   3.0840 -1.3494 -7.6098  3.5232  0.9914
#>  -7.5486  6.9829  2.1340  9.4515 -2.0435
#>  -2.7254  6.1138 -4.2805 -4.4356  2.9620
#> 
#> (1,3,.,.) = 
#>   4.4085  4.7790 -1.9428 -3.0638  3.2076
#>  -9.9792  8.0895  3.3891  1.8191 -7.6854
#>   1.5230  5.6304  5.1435  3.0175 -4.8621
#>   1.6375 -11.4415  0.4565  5.0173  6.4822
#>  -1.1310 -8.5893  3.6073  3.1408  1.6993
#> 
#> (1,4,.,.) = 
#>   5.1920 -2.5063 -3.5074 -1.2248  7.1325
#>  -8.3449 -2.0783 -0.2607  6.3521  7.2339
#>  -1.0204  3.6550  4.0900  2.2109 -1.6304
#>  -6.1780 -2.5103  6.9986  5.4493  3.7610
#>  -5.4089 -0.3607  9.1732 -2.9569  0.7807
#> 
#> (1,5,.,.) = 
#>  -1.2777  1.2476 -5.7416 -5.4315 -2.0373
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]