Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   10.0494   8.3729   0.5716   1.8617  -0.8423
#>   -2.1490   6.6514  -2.6143  -0.0773  -1.5219
#>   -4.6191   2.8416   2.4668 -12.2915   5.5524
#>   -1.9074   2.7545   5.5576  -6.4883  -1.0205
#>    2.5720 -14.1783  -9.1476   5.9686  -0.3635
#> 
#> (1,2,.,.) = 
#>   4.4784  8.3197 -4.2148 -0.8621 -0.4242
#>   4.9106 -2.2142 -9.1411  3.4197 -0.0346
#>  -4.6065  9.0195 -6.3118 -1.0235  1.6422
#>  -6.2746 -6.5941  0.2398 -9.5888  0.9699
#>  -3.5131 -3.2984 -1.6374 -2.5648  3.3479
#> 
#> (1,3,.,.) = 
#>  -2.5625 -3.8717 -10.1279 -1.8689 -3.4950
#>  -0.7339 -6.7959  7.1070  7.3793  3.1919
#>   1.5535  9.3079 -9.0076 -4.0644  3.0567
#>  -4.3543  6.4637  8.0096 -3.8882  4.2477
#>   3.6295 -4.9158  5.7532  4.8834 -4.2929
#> 
#> (1,4,.,.) = 
#>    0.4557  -4.5252  -0.4238   0.4478   1.5038
#>   -1.0329  -2.0716  -3.9553  -3.4665  -0.8101
#>   -1.5354 -17.0727   0.2396   4.7294  -6.2279
#>    7.7439  -3.3353   0.2328   8.9643  -4.4130
#>   -3.3033  10.2295   0.7448   0.5672   0.3617
#> 
#> (1,5,.,.) = 
#>    4.1513   0.3263  -6.2848  -0.6336  -6.3977
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]