Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    1.0499   0.9190   3.0650   2.0074  -3.3131
#>    8.9029   6.9353  -6.3339   1.4508   1.0743
#>    7.2642  13.9998  -5.9891  -3.5572   4.1313
#>   -1.1478   4.6496   0.6966  -5.7598   4.4324
#>    2.9916  -2.7945  -5.2924  -0.4787  -2.3915
#> 
#> (1,2,.,.) = 
#>  -3.5133  0.3886  0.5374 -0.8684 -1.9574
#>   4.4197 -6.3195  0.3462  3.8473 -4.1233
#>   4.0046  1.6794 -3.6364 -5.9407 -2.9276
#>   4.4625 -11.1387  4.1626 -2.6115  0.4287
#>  -0.1334  3.0914  1.5456  5.7432 -3.4407
#> 
#> (1,3,.,.) = 
#>   -0.3495   5.8809  -2.2008   1.9063  -0.5905
#>    1.9351   1.0264   1.8798  -6.0028  -2.6128
#>   -3.7433  15.9050   0.5198   0.6711   1.8251
#>   -4.6576   1.0748  -3.2455 -13.4421  -9.4395
#>    2.8251   8.0224   8.2177   3.1169   5.2293
#> 
#> (1,4,.,.) = 
#>  -2.8535  4.6891  2.3803 -2.0160 -3.1258
#>  -13.6123  2.6031  1.6309 -1.1302 -3.0395
#>   1.6317  3.8311  8.7228  8.5689 -1.2312
#>  -6.6318  7.9549 -6.0399  1.7261  6.2336
#>   0.0274 -3.4343 -3.0917 -5.2981  6.1471
#> 
#> (1,5,.,.) = 
#>   3.8959  1.9680 -2.8729  1.8271 -1.7501
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]