Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -8.5782   0.4528   1.7772  -4.7800  15.1935
#>   -7.2405  -5.1689  -8.6435  -1.7277  -0.7375
#>    0.8208  -3.1727   0.1376  -0.8018  -3.3985
#>    1.0218 -11.4507 -12.4414   0.3190   4.2854
#>   -3.1248   2.0404  -3.9838  -1.0311  -2.2011
#> 
#> (1,2,.,.) = 
#>   -2.2807   1.0608  -1.5452   4.5628   0.3851
#>    0.0717  -0.7261  -2.9149  16.4422  -1.8995
#>   -5.2321   4.9268  -8.9928  -0.8877   4.1658
#>   -2.2479  -0.0051  -8.7791  10.3730   0.8577
#>   -0.7273  -8.5005   7.2164   1.9553   1.9309
#> 
#> (1,3,.,.) = 
#>   3.7869 -4.8477 -0.5042 -3.1657 -0.4664
#>   1.2527 -6.7604 -8.6982  4.3850 -1.1723
#>  -3.0847 -2.6301 -4.5965 -2.7329  6.0259
#>  -0.4475 -9.2790 -5.1431 -0.1351 -6.2608
#>  -2.6604 -2.8576 -7.7808 -2.7451 -0.2996
#> 
#> (1,4,.,.) = 
#>   -2.7858   1.2784   8.4755   0.8621  -0.6680
#>   13.2627   4.4286 -14.9865   2.9841  -1.7695
#>   -0.8464   7.7995   5.3040  -4.4092   4.9925
#>    2.9660   0.3360  -1.4193  -3.3490  -4.1547
#>    3.5958   2.7023   6.7268   2.9143  -1.4168
#> 
#> (1,5,.,.) = 
#>   2.3830  3.4695 -8.0905  1.0061 -0.1861
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]