Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -1.8342  -2.4572  -2.3684   4.7146  -3.2447
#>    3.8226   1.2514   3.0124  -0.7052  -9.2219
#>   11.0261  -5.0516   4.5049  -5.8000  -1.1283
#>   -8.5793  -2.4472   5.2401  -3.0461   1.3591
#>   -2.2965  -3.5319  -5.1625   0.2653  -2.6202
#> 
#> (1,2,.,.) = 
#>  -3.1008 -0.1081  5.2397 -2.3117  1.2218
#>   1.2318  0.4446 -1.3858  1.2707  1.2250
#>   1.5256  8.1783  2.2618 -3.3660 -1.3920
#>   0.4458 -5.3380  6.3956 -8.2323  3.3141
#>   3.6885  5.9161 -0.8618 -9.4669  1.8891
#> 
#> (1,3,.,.) = 
#>    7.3113  -2.2619   0.0596  -6.1693  -0.7649
#>   -2.8382  -6.7717  -3.1046  -0.5185  -5.1656
#>   -4.2752   3.6677  -2.2686   4.0481  -5.4733
#>    1.6219   5.7588  15.3677  -1.4419   0.0892
#>    4.2582  -9.8247   3.7307  -4.7627  -4.5579
#> 
#> (1,4,.,.) = 
#>  -3.0661  0.8255 -0.6500  4.0684 -2.0807
#>   4.7517  4.2365  4.2172 -5.3061 -2.9095
#>   2.0619 -3.6984 -0.4676  7.0817 -11.3288
#>   6.8238 -8.7271 -5.8662  0.5398  0.2077
#>   5.0487  3.5289 -0.0304  8.3216 -5.7129
#> 
#> (1,5,.,.) = 
#>    3.6939   3.1262  -1.1671   0.2329  -0.7270
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]