Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -2.2092  -1.9330  12.5686   0.3296  -0.9045
#>   -9.9111  -4.9568   9.9311  -0.5410  -0.3247
#>   -2.3690  -1.1269  -2.7990   1.0996  13.2528
#>   14.3730  -7.3771 -13.7455   2.0831   1.0651
#>    9.1954  -4.3643   2.1820  -2.9379  -4.3166
#> 
#> (1,2,.,.) = 
#>    0.9449  -2.0998   5.7115  -3.8022  -4.6403
#>    7.2922   0.8877  -7.6691   5.1574   9.5829
#>   -3.4020  10.9590   7.5524   6.7212   7.7795
#>   -5.3269  -6.3287   8.5487   2.0058  -6.2504
#>   -3.9781  -1.6485  -4.1315   5.1656  -1.8991
#> 
#> (1,3,.,.) = 
#>    9.0872   7.6925  -3.4425  -4.5503  -5.8537
#>   -1.4785   7.1734   5.7936  -5.6837  -5.3265
#>    5.3660  -1.6547  -3.2874   0.7544  -4.7628
#>   -1.7339  -1.3810  15.5370  -1.4329   3.8254
#>   -5.8451 -11.4032   1.5171  -5.2256  -3.1182
#> 
#> (1,4,.,.) = 
#>  -1.6069 -1.5657 -2.3840  6.8279  6.2280
#>  -7.3543 -1.5658  4.7794  6.8587  1.5457
#>   0.4776 -6.7385  6.1378  2.1063  3.0263
#>   0.8422 -5.1849  1.4116 -6.4876 -0.6184
#>   2.8276  0.2370 -3.9376 -10.2066  0.2551
#> 
#> (1,5,.,.) = 
#>  -9.5507 -9.9299 -0.1540  0.1199  2.1469
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]