Conv2d

Conv2d

Arguments

input
NA input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
NA filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
NA optional bias tensor of shape \((\mbox{out\_channels})\). Default: None

stride
NA the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
NA implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
NA the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
NA split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=None, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See ~torch.nn.Conv2d for details and output shape.

.. include:: cudnn_deterministic.rst

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   12.7403   3.1743   4.2664  -1.1589  -4.7747
#>    4.3204  16.0023   2.9855  -0.4251  -4.5754
#>   -3.7396   0.1155   7.2139   0.0566   0.1682
#>    8.0297  -3.6696  -5.3214  -2.1186  -2.7683
#>   -4.2373   4.9410  -3.4199   1.3600  -1.9029
#> 
#> (1,2,.,.) = 
#>  -0.8476  6.0065 -2.5265 -5.4223  3.5681
#>   5.4564 -0.6011 -2.1892 -0.8665  3.2542
#>  -6.8643  5.8655  0.1337  3.1428  1.9489
#>   3.5146  8.7320  6.8901  6.5349  3.3823
#>  -1.5378 -8.0291 -2.4452 -2.1523 -2.1974
#> 
#> (1,3,.,.) = 
#>    1.5959  -3.6727   5.2818  -9.9311  -3.2240
#>    6.7801   1.5306   1.9910  -2.6695   5.1391
#>   -1.3441  -0.8893   3.8828   2.8139  -6.0091
#>    1.1648  -3.7450   5.8186  -2.3897   1.3658
#>    5.0413  10.6879  -2.4910   4.1133  -3.0309
#> 
#> (1,4,.,.) = 
#>    5.1706  -7.3330   2.3214  -0.8919  -1.7919
#>   -1.1581  -0.9756   3.1245   4.7576  -1.1151
#>    0.5424  10.0845   2.0282  11.4451  -1.5601
#>   -0.9402 -11.5202  -4.3089  -3.0170  -4.0573
#>    5.8633  -1.2902  -0.2532   0.7452  -0.1586
#> 
#> (1,5,.,.) = 
#>   2.3595  1.7426  5.1700 -0.7004  4.1723
#>   1.0678 -4.0691 -6.3489 -5.8262 -1.2365
#>  -4.3847  7.4245  1.3517 -4.4424  1.6145
#>   9.5457 -2.1841  5.2564  2.3278  3.8379
#>  -5.0946 -3.1453 -5.1444  2.0733  1.2251
#> 
#> (1,6,.,.) = 
#>   -4.2687  -2.7930   4.4640  -9.2532  -4.9388
#>   10.6985   9.1275  -0.2807   6.3323   0.0911
#>   -4.9260   0.3375  -8.6046   0.5057  -4.7161
#>    3.6786  -5.1994  10.8921   3.9161   6.0740
#>    3.6582  -7.0904  -2.1222  -5.4795  -2.4925
#> 
#> (1,7,.,.) = 
#>  -2.8982  0.6261  1.2095 -5.0351 -0.8326
#>   5.1963  1.6991 -4.0706 -0.5831  3.5483
#>  -4.7321 -3.1141 -3.5572 -6.1051  3.5246
#>   3.9899  0.0364 -2.0801 -2.9640  1.9790
#>  -4.9582 -5.8264 -3.9300 -3.3669 -2.7127
#> 
#> (1,8,.,.) = 
#>    2.1049  -4.5380  -1.8678  -4.7371  -4.7240
#>    9.0626  -3.4540  13.2130  -5.9710   2.0413
#>    4.5687  -2.4184   0.8724  -1.8230  -1.1643
#>    2.6529  10.6865   7.4968   5.4103  -3.1017
#>    2.6947  -1.5496   5.1085  -3.1549   1.4289
#> [ CPUFloatType{1,8,5,5} ]