Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -1.3819  2.9029 -0.0468 -6.4783  2.4554
#>   0.7238  1.7946 -0.7321 -1.7861  0.4766
#>  -0.7696 -4.6541 -3.1567  5.6900  1.1359
#>  -7.5980  3.8347  0.2774 -6.7750  6.2029
#>   4.5506  4.3363  6.1995 -1.9137 -4.3008
#> 
#> (1,2,.,.) = 
#>  -3.8518  2.9720 -4.4454 -3.2598  7.8787
#>   1.3998 -3.0209  2.5758 -0.4770 -0.4357
#>   1.5303  3.2127  4.1271 -10.5246 -4.1831
#>  -2.2720 -4.3406  3.9584 -9.0855  7.4150
#>  -3.9973  3.9840  9.7885 -4.4131  1.3867
#> 
#> (1,3,.,.) = 
#>   -5.8712   6.0248  -3.5324  -1.6031   2.2199
#>   -0.2201  -4.8783  -5.2249  -4.0542   3.0859
#>    0.1684  -7.2219   0.7955  20.8519  -0.5232
#>    4.8606   7.6706   0.0596  -8.3861   4.8869
#>    4.2143  -2.8154  -0.1388  -7.4026   1.2008
#> 
#> (1,4,.,.) = 
#>  -5.4789 -1.9614 -2.8898  1.8500  0.8066
#>   3.2866  3.9059 -11.0890  9.0730  3.6360
#>  -1.5545 -1.1816  1.2695  2.0842 -0.4859
#>   5.8627  9.2832  3.2062  7.7787 -8.9146
#>   0.2396 -10.4538  1.3707 -7.3728  5.0564
#> 
#> (1,5,.,.) = 
#>  -0.6298 -3.2100 -2.1412 -2.1606 -2.5763
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]