Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    1.7254   1.5360  -6.9906   0.0486  -1.9488
#>   -4.3483  -5.9518  -1.0718  -2.0334   1.1878
#>   -0.7058   1.0040   7.2343  10.3486  -0.2613
#>    1.3598   1.7507  -1.0039  -0.0228  -3.2656
#>   -2.1291  -4.0008   4.4476   1.1822  -0.0014
#> 
#> (1,2,.,.) = 
#>   -9.4805  -0.4300  -7.6604   6.0327   4.2416
#>    0.6938   0.1908  12.7536  -9.9278  -5.0835
#>    4.7813  -0.8086  -7.4148  -0.5424  -8.4933
#>    1.9637   8.3088  -5.8763   1.5336   7.1313
#>    4.6965   5.7377  -1.4127 -11.7435   4.7713
#> 
#> (1,3,.,.) = 
#>   2.8252  0.2338  7.5793 -6.2875  2.1725
#>   2.7578  1.2738  4.2520  1.8493  0.7565
#>   1.3741  6.3664 -7.0633  3.2830 -0.1154
#>  -2.9411  1.0252 -4.7462 -6.9356  6.9906
#>  -1.1749 -3.7680 -0.2200 -3.3196 -1.4958
#> 
#> (1,4,.,.) = 
#>   -3.2850   8.8332  -2.7964  -7.2271  -5.9129
#>    3.9732  -4.3375   2.6295  10.7297   0.1509
#>    5.8635   3.9316   6.0393  -8.2306   1.2365
#>   -1.6016   5.3287 -11.5654  -1.5202  -4.0841
#>    0.9283  -1.9636   7.9763   0.5117  -2.5095
#> 
#> (1,5,.,.) = 
#>    2.9465  -0.3805  -0.6042   1.1483   3.9806
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]