Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -3.5111  -2.7042  -6.3748  -3.6516  -2.5321
#>   -5.9368  -5.2648   2.2961  -2.2259   4.5112
#>    0.5910   3.0160   9.5183   4.5512  -0.4351
#>   -5.2675  -3.5527  11.6300   3.5522  -3.3478
#>   -2.0893  -4.8931   5.0405   5.1630  -4.8255
#> 
#> (1,2,.,.) = 
#>   2.6551  1.0520 -3.6052 -3.1616 -6.2554
#>   5.0609 -1.0506  0.3881 -2.0105 -4.4277
#>   4.6707  0.3970 -1.7106  4.9149  2.8757
#>   3.1563  0.7800 -2.4278  7.1315 -3.3621
#>  -1.9816  0.9222 -1.9924  1.6961  4.6556
#> 
#> (1,3,.,.) = 
#>   2.4998  2.9237 -4.2844 -1.2551 -6.6879
#>  -4.0018 -7.5700 -7.8253 -10.2765 -1.2123
#>  -3.7866 -3.9454 -3.6632 -0.5074 -6.2853
#>  -6.6963 -6.9882 -6.5469 -10.0054 -3.1180
#>  -9.4356 -7.6140 -2.1418  1.6214  2.6935
#> 
#> (1,4,.,.) = 
#>   -6.7099   6.9575   8.3067   3.6028  -0.2321
#>    6.6693   2.9523  -0.2451   0.0231  -6.0445
#>   14.1405  -9.2593  -8.7378  -9.6682  -8.8534
#>    9.5104  -7.4808 -11.2558  -3.3226  -2.9364
#>    6.9412   3.3740  -5.2980  -1.4779  -3.3840
#> 
#> (1,5,.,.) = 
#>    1.8978  -1.1075   3.2545   4.2548   7.7413
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]