Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    2.3082  -7.4239   6.8365  -5.6187  -1.8196
#>   -1.0045   6.6026  -0.5694   4.3607  -0.8080
#>   10.4554 -11.5170  13.6917   2.7655   1.1333
#>    2.8588  -1.6884  -0.7491   0.7908   0.9478
#>   -0.8593  -1.5077   3.2329 -10.8562  -1.7578
#> 
#> (1,2,.,.) = 
#>  -0.4726  4.8031 -6.8435  5.1651 -1.7856
#>   4.1196 -8.0434 -6.1372  8.9356 -1.1644
#>   1.6149 -3.0136 -11.0907 -8.3511  3.4666
#>  -0.9003  4.3284  2.7067 -6.5086 -0.8184
#>  -2.4514 -7.7530  0.5918  2.4890  5.5416
#> 
#> (1,3,.,.) = 
#>   -0.9957  -5.6880   3.0121  -1.9031   0.4371
#>   -2.4791   1.8003 -10.8587   3.0963  -1.5514
#>    0.3178  10.4734  -4.5888   2.2976  -5.1962
#>    1.3398  -5.9335   1.5559  10.2611  -7.8015
#>   -5.3643   1.8930  -6.8060   4.3596   7.7495
#> 
#> (1,4,.,.) = 
#>    1.1316  -5.7452   4.8674   1.9472   0.8832
#>    3.3714   4.3976  -4.7114   1.8798   1.4471
#>   12.4108  -5.0290   3.5966  -3.5025  -2.6582
#>    4.2454 -10.2660  -4.2489  -0.8993  -5.5608
#>   -0.7154  -3.0817   8.0348  -6.7649  -0.4081
#> 
#> (1,5,.,.) = 
#>  -1.2216 -1.6101  2.2096 -4.9992  0.2219
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]