Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    0.0601   2.4654  -1.5006  -0.2293  -0.2000
#>   -3.4628   3.5616  -3.5736  -3.2882  -3.2509
#>   -3.1711   0.8777   4.2728   3.5395   4.6853
#>    1.4066   2.3239  -6.7979  13.3377  13.6287
#>    6.1164   0.6429  -3.3886   1.1149   2.4793
#> 
#> (1,2,.,.) = 
#>   -2.4497   0.3328  -1.5141  -2.4401  -4.8773
#>    1.2175  -3.5088  -9.6157  -8.0809   1.7349
#>    0.1354  -6.5738  -9.2620  15.9911   3.8140
#>   -4.7238  -2.3463  14.0089  -4.4159  -0.9123
#>   -3.2717  -9.4104  -6.7307   2.2746   5.1673
#> 
#> (1,3,.,.) = 
#>   -4.2834  -4.3024  -0.9671   3.2942   1.4540
#>   -8.6346   4.7586  -2.2132  -2.3384   1.6963
#>   -1.8277   0.4283  -8.6707   3.6723  12.4993
#>   -4.8293   3.1942   8.3663   5.3767  -6.8951
#>   -4.2390   2.4615  -1.7633  -2.9671  -3.4298
#> 
#> (1,4,.,.) = 
#>    1.9470   0.4786   3.4822 -10.5924   1.6014
#>   -2.9517  -6.0323   5.8886  -7.1422  -1.4755
#>   -7.7668   3.6245   8.1834   2.5265   4.4149
#>    2.3928  13.2671   1.9446  -3.1121   3.4607
#>   -2.2584  -0.2200  10.7155  -0.8928  -2.7484
#> 
#> (1,5,.,.) = 
#>   0.4049 -1.6191 -0.1799 -4.0668 -3.7067
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]