Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   4.4970  3.4864 -0.6793 -1.0731  2.1858
#>   7.0990  3.7082  6.5866 -2.8806 -2.8763
#>   1.3625  2.6945  1.4323 -9.5622 -2.9163
#>  -2.9226  2.8969 -6.3858  9.2082 -3.0913
#>  -4.5888 -0.2497 -5.6967  3.3050 -5.1690
#> 
#> (1,2,.,.) = 
#>  -2.3636  0.3488  0.9187 -0.7605 -4.9159
#>   0.9268 -7.1368  7.4800 -8.6739 -1.0022
#>  -2.2843 -2.3828  6.5002  3.1360  1.2008
#>  -1.5604  4.1455 -6.4348  5.2361 -5.9527
#>   5.6411 -4.1684 -0.4588  0.3485  0.6383
#> 
#> (1,3,.,.) = 
#>  -4.4903 -2.3423 -0.0651  1.2908 -2.8397
#>  -1.2935 -1.5900  5.6460  0.6256 -9.9700
#>  -5.5029 -3.1872 -9.7560  6.1097  3.4764
#>   2.9966 -5.5509  0.5144  1.6999  2.8689
#>   6.3662 -2.7968 -8.6420  2.0845  1.0004
#> 
#> (1,4,.,.) = 
#>    0.4214  -4.6839   3.3181   6.8284   6.9999
#>   -7.3308  -4.7744   0.7230   4.6826  -0.8534
#>    0.3276  -1.9913   4.5275  -0.5761   8.9290
#>    5.9225  -6.3257  13.8425   8.8808   9.3700
#>    2.9453  -2.1341  -1.4374   1.9609   2.8834
#> 
#> (1,5,.,.) = 
#>   -1.8847   2.7779  11.1969   8.4794   6.6241
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]