Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   3.1462 -5.6672 -7.9740  2.0895 -5.8411
#>  -3.9699 -3.5066 -3.7618 -11.4777  0.1851
#>   3.5365 -0.9169 -1.5443 -0.1142  5.3145
#>  -2.2553  4.2938  0.8427 -1.7506  2.1276
#>  -5.6569 -0.9172 -3.8000 -0.1372 -2.4545
#> 
#> (1,2,.,.) = 
#>   -1.5387  11.4528   0.0857  13.4854  -0.0069
#>    1.1582 -10.5980   7.0229   3.7672   3.1583
#>    5.2189  -1.5535  -4.8103   3.5332  -5.9241
#>   10.7660  -2.0549  -0.4894   7.2897   2.1240
#>   -4.9937   5.3330   1.8200   0.0201   1.9035
#> 
#> (1,3,.,.) = 
#>  -0.5344 -7.1500 -2.1815 -6.4052 -2.5355
#>  -0.9795 -1.1095 -7.7243 -8.9257 -6.4259
#>  -4.1859 -1.6798 -2.4022  1.4776 -1.5895
#>  -7.8867 -4.8088 -3.5644 -5.9332  0.8315
#>   1.1898 -4.0646 -0.3369 -3.1258 -6.4168
#> 
#> (1,4,.,.) = 
#>  -4.2186 -2.2330 -0.2732 -9.0199  2.3865
#>  -2.8896 -9.1167 -4.1611 -2.7136  0.6530
#>  -4.9429  0.5739 -1.5872  0.8260  1.5845
#>  -1.1030 -0.7287 -16.7118  2.8276  0.0362
#>  -5.3900  0.9149 -3.1601 -7.8542 -0.7762
#> 
#> (1,5,.,.) = 
#>  -4.7852  0.1323 -0.1621 -2.1654  2.2967
#>  -5.0110 -6.8604 -2.5183 -0.7330  2.0924
#>  -1.1189 -3.3173 -7.1832 -3.9109 -3.8811
#>   1.6551  2.3243 -0.9795 -3.0309  3.2335
#>   3.0468 -1.2222 -5.1248  2.3748 -3.3063
#> 
#> (1,6,.,.) = 
#>    6.1559   6.9945  -4.0617  -0.6720  -2.4359
#>   -3.1206   1.2034  14.5700  -5.1990   6.1557
#>    1.6713  -4.9825   0.2442  -5.9373  -0.2142
#>   -0.3297   1.1878  -4.5510  -0.5501  -8.2094
#>    1.5988   7.2115   1.7430   5.3158   5.0985
#> 
#> (1,7,.,.) = 
#>  -0.9358  4.7896 -6.3822  2.0359 -5.4548
#>  -4.1794 -5.1443  7.6380 -2.7219 -3.2543
#>  -1.9499  0.8026  1.4290 -0.3331  0.7852
#>   4.0708 -6.7176  4.8772  4.4655 -1.8020
#>  -1.9156  3.3183 -1.7985  5.0085  1.2032
#> 
#> (1,8,.,.) = 
#>   4.9318 -2.1501  1.1723  7.7414  3.1249
#>  -2.7583 -12.7894 -10.9663  4.3047 -1.1796
#>   4.3779  1.8231  0.0528 -3.4288 -0.3644
#>  -3.3108 -3.7445 -6.9456 -6.8718  7.3775
#>  -9.2444 -4.6305  6.0062 -3.1374 -5.1606
#> [ CPUFloatType{1,8,5,5} ]