Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -1.0569  1.5447 -3.0220  4.9441 -4.6343
#>   0.3067  0.2752  3.5238 -3.2304 -2.1347
#>   4.6321  4.7693  6.9399  7.3575  1.1961
#>   2.4534 -6.5831  2.1187  2.4684 -1.9805
#>   3.6331  6.0727 -2.2551 -1.6256 -0.4263
#> 
#> (1,2,.,.) = 
#>   -2.3652  10.3973   8.8094   2.9988   8.4687
#>  -13.1726  -2.0607  -3.9400  -4.0538   0.2729
#>    3.7647  -9.4965  -6.2523  -4.8035  -1.2289
#>    1.7356  10.7615   9.1931  13.3766   8.0847
#>    4.6093  -1.7724  -7.2234  -6.6927   1.9376
#> 
#> (1,3,.,.) = 
#>   -0.4252   0.2073  -7.2160   4.2250   0.1889
#>   -2.2295   6.4681  -4.4556   7.0570  -6.3076
#>   -1.2020  11.7436   1.3505   1.3981   3.2623
#>   -5.0591  -0.2234  -4.2104   4.0892   0.0219
#>    1.0978  -5.2579  -6.3781  -8.9831   5.3737
#> 
#> (1,4,.,.) = 
#>   -7.7089  12.2909  -2.3906   7.3614   1.4749
#>   -4.1968  -7.6225  -2.3947  -1.0748  -0.5053
#>   -0.9711   7.5685   3.9519  -6.1794  -1.3409
#>    9.0729  -8.4759  -9.3343  14.6863   5.5061
#>    2.1274   6.2149  11.3055  -2.3626   2.4173
#> 
#> (1,5,.,.) = 
#>   5.3125 -1.8211  4.9858 -5.1688 -0.9161
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]