Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -3.5314 -1.8066  4.0618 -3.4512  1.3936
#>   2.8689 -3.8876 -3.7471 -4.5023  0.6428
#>  -3.8486 -0.0600  6.9223 -6.0743  8.4950
#>   6.8957  4.3490 -10.4008  8.5298 -4.4190
#>   4.8391  3.8955  5.9504 -2.4804  6.2833
#> 
#> (1,2,.,.) = 
#>   -4.5897  -4.4056   0.5333  -1.8255   7.8957
#>   -5.7344  11.3735  10.2611   2.6200  -7.7254
#>    2.0461   0.1741  -1.2510 -10.5765  -3.3641
#>   -4.5825  -3.9558   4.5270   0.1919  -2.2728
#>    1.9727   4.7667 -20.9282   4.7403   4.6348
#> 
#> (1,3,.,.) = 
#>  -0.1141 -5.0596 -5.6699  3.3055 -2.7024
#>  -7.8050 -10.7413  0.8281  9.8138 -2.3302
#>  -0.5440 -4.4533 -0.5605 -11.0678  3.2417
#>  -0.0695  4.6700 -0.3693  2.4681  2.2692
#>  -0.4375  7.1192  1.9826 -9.5759 -4.8570
#> 
#> (1,4,.,.) = 
#>    3.0798   8.4109  -8.1660  10.3865  -8.2667
#>    6.5151  -4.0383   6.0067   5.8623   1.7767
#>    0.9888  -2.2142   5.4664  -8.6149   1.2578
#>    1.5753   2.2107 -10.7369   1.0911  11.3145
#>    3.5959   2.6298   1.8454   3.6432   4.4221
#> 
#> (1,5,.,.) = 
#>   -2.0922   5.2017  -1.4702   4.6054  -7.7932
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]