Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    1.2852   7.5605  -3.1620   2.6189  -7.4122
#>    3.8719   1.9485 -12.0423  -6.0115 -14.4482
#>    1.8714   1.8002  -0.8304  13.2517   3.4023
#>    3.8953  -5.3520  -8.2082 -15.3719  -4.6276
#>   -7.9871   3.5364   0.7894   7.8183   7.5267
#> 
#> (1,2,.,.) = 
#>   0.7501 -0.2025  0.2154  6.8863  5.0910
#>   1.5924  7.6651  5.4175  4.8588  5.0483
#>  -2.7835  9.5132  7.0208  6.6230  4.4685
#>  -0.1493  2.9963 -6.6257  0.8897  1.7003
#>  -2.4292  4.2030  0.4032  4.0788  4.0859
#> 
#> (1,3,.,.) = 
#>    0.9802   1.5412   2.5753  -8.3827  -0.7817
#>   -5.5264  -0.5652  -0.6741  -9.5172   7.1923
#>   -0.9164   2.4476  -5.0768   6.2129  -1.5872
#>   -3.4612 -12.1874  13.2705  -1.1590   2.8937
#>   -7.2189  -8.3360   1.9999  -1.9636  -7.2850
#> 
#> (1,4,.,.) = 
#>   -0.5423  -3.5813   5.4783   5.5420   5.5022
#>    6.4560  14.0806   1.9536   6.5103   2.2300
#>   12.3066   5.8362   1.3354   1.1963   1.1751
#>    3.3102   5.9839  -4.5719  -4.6628   4.1099
#>   -1.5655   1.2670  -4.4087   1.6273   2.0346
#> 
#> (1,5,.,.) = 
#>    3.0838  -0.0024  -0.4908   3.5461   1.0776
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]