Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   4.0243  3.7376 -3.1911 -4.8255 -6.8137
#>  -2.1746 -14.9573 -3.8870 -5.3694 -6.0203
#>   5.5052  6.6188 -5.0687  9.0323  0.4659
#>   7.8399  3.9780  6.0283  1.4553 -6.9449
#>  -2.1688  0.1157 -0.0550 -3.0811  4.6206
#> 
#> (1,2,.,.) = 
#>   -3.6353  -1.5130  -2.6996  -3.7135   5.5139
#>   -1.2883   3.2130   7.1810  16.2683   8.0086
#>   -7.4226   9.2379   7.1055   1.8173   2.6065
#>    0.1806   6.4647  -6.3935 -10.6666 -11.5207
#>    1.5135  -6.2188  -7.0346   3.2294   2.3692
#> 
#> (1,3,.,.) = 
#>   -4.3837  -7.0316   6.2504  -2.9672  -6.1117
#>   -8.2633   0.5069  -7.1323 -12.6358  -4.5621
#>    3.8777  -1.6151   0.7176   5.8723   5.1316
#>   -3.1265   8.1339  12.4319  14.7431   4.1683
#>   -3.2443  -1.5852   2.8617  -4.7472  -0.8379
#> 
#> (1,4,.,.) = 
#>  -0.0548 -6.2184 -0.2030  2.8197 -3.1074
#>   5.8858  5.1368 -0.4416 -3.3813 -0.2747
#>   9.2711  2.5429 -11.5141  2.5735 -0.2128
#>   0.9522 -4.5155 -8.3986 -3.7373 -6.8189
#>  -2.5948  6.7073  0.9943 -4.4844  1.2209
#> 
#> (1,5,.,.) = 
#>   -5.1566  -0.7991   4.1772   5.3517   2.2563
#>   10.1776   2.6731  -3.6359   6.6869   5.8780
#>    2.4127  -1.5569  -0.1435  11.2920   4.1110
#>    8.6535   4.2422  -1.7197  -1.9647  -4.0549
#>    8.1580   7.3564  -5.6078  -4.1503   3.3268
#> 
#> (1,6,.,.) = 
#>   1.4464  2.1223 -0.0925 -4.2614  1.7261
#>  -1.5468  8.1299 -1.7983  6.8752  5.8608
#>  -0.5857 -5.6725  1.9944  2.4742 -3.5909
#>  -2.3867  2.1849  4.0331 -2.2902 -6.3801
#>   3.9584 -2.5554 -1.1067  0.9581 -2.4002
#> 
#> (1,7,.,.) = 
#>   3.0118 -2.6042  0.2651  2.5277  0.0028
#>  -3.3712  1.7879  0.9126  5.0013 -3.2194
#>   8.9812 -3.9752 -3.5089 -8.9629 -12.7984
#>  -2.4382 -6.5683 -9.0386 -1.0824 -0.8877
#>  -0.0665 -1.3189  1.3501  4.3611 -1.4619
#> 
#> (1,8,.,.) = 
#>   -1.8375  -5.1074  -9.7494  -8.4274   0.7705
#>    3.3095   0.7931  -2.0577   3.3355   0.5141
#>   -2.3046  -2.6208   7.0623   2.9831   4.5503
#>    3.1643  12.0367   5.1488  -5.1226  -9.9351
#>   -3.0092   5.9618  -0.9362  -5.6725   0.3272
#> [ CPUFloatType{1,8,5,5} ]