Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    4.2777  -3.6893   5.7045  12.5031  -6.9851
#>    3.4780 -14.3218  14.4795   4.0467 -12.3221
#>   -9.3499  -5.3087   1.0127   7.7204  -6.1998
#>    6.6114  -3.9279 -12.3504   1.7439   4.0547
#>    5.9614  -5.2096  -4.0449   4.5764   4.5578
#> 
#> (1,2,.,.) = 
#>  -1.3833 -3.1494  2.4787 -0.5415  0.7495
#>   2.8318  0.4295 -4.3451 -4.5279  5.2137
#>   2.7529  5.4469 -2.2190  1.6766 -0.7859
#>   6.1252 -2.1602 -4.9213 -3.6542 -0.0678
#>  -2.3520 -8.2281  1.6783 -3.3883 -1.5895
#> 
#> (1,3,.,.) = 
#>   -0.1564  -6.2336  -3.4995  -5.3862  -6.9332
#>    1.2718   1.7480  14.2907   4.4857  -2.5967
#>    3.3010  -2.1314   4.2473   1.7082   3.6511
#>    3.8269   1.8548   0.3609  -3.6272   2.9850
#>    4.1578   1.6844   3.1144   8.5628  -3.7615
#> 
#> (1,4,.,.) = 
#>  -3.3488 -11.6307  1.5314  1.6313  2.6113
#>  -6.1197  6.2264  4.7419  0.8775  7.1474
#>  -1.6415 -5.4413 -10.7845  6.0031  5.1334
#>   1.6471 -8.7156 -2.6459  8.8837  4.7778
#>   0.6365 -5.5838  0.9536 -1.0834 -3.2699
#> 
#> (1,5,.,.) = 
#>   0.4182  5.4167 -3.0617 -10.5768 -1.3508
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]