Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -1.2072   1.9123   0.8207   4.7609   4.7965
#>   -2.8802  -8.2586   3.2458   1.6755   3.4293
#>    4.4970 -14.7177  22.7660   9.2153  -8.8881
#>   -2.7035   2.1788  -1.6816  -5.3574   3.2081
#>    0.2316   7.7389   1.2523  -5.3880  -5.1406
#> 
#> (1,2,.,.) = 
#>  -1.2491  1.3822  1.3456  2.6971 -5.8119
#>   2.0908  3.9267 -3.7326  8.1962  3.6213
#>  -0.8375 -6.6431  8.5051  8.8476 -8.4967
#>  -9.6882  5.1370 -13.1065  2.0940  0.0636
#>   3.3151  1.4656  0.3896 -3.3033 -5.6563
#> 
#> (1,3,.,.) = 
#>  -3.7594 -7.9765 -2.9085  5.1037 -8.3409
#>  -6.5386  7.7398 -9.2263  3.7157 -5.8678
#>  -7.8296  4.8639 -10.1191  2.8080 -3.1094
#>   4.7727 -12.4309  7.0908  0.6128 -2.5439
#>   2.3913 -2.5282 -1.1988  0.5432  2.7851
#> 
#> (1,4,.,.) = 
#>  -4.7977 -8.2501 -3.5003 -3.6505 -7.9129
#>  -5.2363 -4.2171 -4.8737  2.1883 -6.4016
#>  -5.6949  4.5269  1.1945  4.9270 -8.8006
#>  -0.6048  7.0043 -1.7695  1.0189 -2.0242
#>  -3.4895  3.5810  0.6347 -7.2982  2.8114
#> 
#> (1,5,.,.) = 
#>  -1.8557  2.5078  5.4092  5.7864  3.9220
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]