Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -4.9953  3.0351  4.1265  2.5099 -2.7899
#>   3.5744 -7.5846 -4.5152 -0.2476 -2.9935
#>  -7.0376  6.6112 -0.6818  0.3155 -6.4951
#>   7.0606  4.7149  5.6289  5.9330  1.2233
#>   4.4133 -7.4515 -1.3810  5.4669  0.3031
#> 
#> (1,2,.,.) = 
#>  -0.7499  0.8187 -1.1258  2.8162 -2.2980
#>  -8.2486  0.1606 -0.6782  0.8963 -11.7464
#>  -2.2378  5.2654  5.4524 -1.4058  1.2655
#>   1.5492 -0.6582  3.9603  8.2172 -9.3447
#>   1.5390 -0.2889  1.6293  1.3816 -2.9867
#> 
#> (1,3,.,.) = 
#>   1.6314 -1.9375 -7.7929 -7.1623 -0.5022
#>  -9.5436 -3.3903 -2.8160  7.5514 -5.2835
#>  -6.5327 -6.2378  5.3714 -0.3869  9.0308
#>   9.8240  1.2959  4.6514  0.4207 -1.9778
#>  -6.8178  3.5944 -7.1679  4.4383  2.1931
#> 
#> (1,4,.,.) = 
#>    1.7755   3.2979   1.5824   2.3718  -4.0408
#>    7.1193  -4.5467  10.1893  -9.9998   2.6714
#>   -4.3475  -7.3542   4.5670  -1.6060   6.6152
#>   -5.2217  10.8243  -2.0110 -12.9524  -3.1391
#>   -0.0179  -4.4946  -4.3297   3.1157   0.6201
#> 
#> (1,5,.,.) = 
#>    7.1808  13.9802   4.5304   7.8096  -1.6655
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]