Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    3.1905   0.7080   1.9172   9.2368  -6.5454
#>    0.8438  -2.3981  -2.4745   3.2485  -8.2714
#>   -0.5787  -1.5082   7.1595   8.1050   9.6473
#>    1.1697   7.1012  20.2860  10.4228  -2.5828
#>    1.4599   4.1446  -1.6478  -1.9676  -3.1971
#> 
#> (1,2,.,.) = 
#>    0.3065   6.6891  12.1709  -1.1086  -1.2834
#>    3.5097   2.7329   5.7062   0.2898  -0.2819
#>    5.8836   1.9909   4.3501  -7.0390 -10.0283
#>   -0.4801   2.1408  -6.4691  -7.1385  -3.3886
#>    5.3276  -5.9525 -12.1445  -4.1106  -2.8633
#> 
#> (1,3,.,.) = 
#>    2.3491  -0.6614  -0.0438 -11.5443   0.4086
#>   -5.1060  -2.4466  -2.0663   5.2575   8.2057
#>   -6.5829  10.9107   6.1101  -3.0145  -0.6999
#>   -2.0586  -8.6677  -3.1851  -5.3759   2.4746
#>    0.8905  -6.7834   6.0525   6.2993   1.4434
#> 
#> (1,4,.,.) = 
#>   -1.0859   2.2116   2.7386  -0.5821  -5.0712
#>   -7.2477   2.3523   8.0329 -12.3038   1.7497
#>   -5.5950   3.2315  -0.7942   2.9327   8.6853
#>   -0.7146   4.6161   7.7387   5.7248   1.0504
#>   -0.3697   1.9362  11.4641  11.3247   6.9968
#> 
#> (1,5,.,.) = 
#>    5.0300   1.2735  -6.2140   5.7788 -11.4515
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]