Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -5.1931  1.4041  4.1435  3.8637  0.9316
#>  -3.0061 -0.5271  3.9516 -4.5792 -1.8635
#>  -6.4293 -8.3294  8.2632  7.9648  4.5316
#>   7.7539 -8.4685 -2.3763  1.1200  6.2595
#>   1.3047 -6.1807 -3.7363  5.9323 -0.8288
#> 
#> (1,2,.,.) = 
#>    5.9928  -1.6809  -0.5438  -0.7543  -2.2308
#>    5.9829   0.1353  -3.5862 -10.7259  -4.4990
#>   -6.7754  12.6271   7.0938   9.3049 -10.6004
#>    4.3933  -9.5856 -13.4671  13.7459  -4.3085
#>    2.9043  -0.7294  -9.1553   3.1174  -6.3283
#> 
#> (1,3,.,.) = 
#>   -5.2144   3.2791  -1.3664   0.8487  -0.1552
#>   -0.0203  -0.0495  -7.7625  -8.0313  -5.5007
#>    3.9195   6.6020   7.6019  -8.8771   6.8147
#>   10.7026   4.8836  -4.6124 -10.2992   1.3530
#>   -2.0601   3.4591   9.4001  -6.3832   2.3264
#> 
#> (1,4,.,.) = 
#>   1.1674 -3.1261 -1.8810  5.0503  2.9096
#>  -1.1782  0.5891  0.4977 -4.0540  3.0384
#>   2.0771 -6.7067  0.4256 -4.5138 -0.2659
#>  -4.4625  0.9335  9.4598  4.7847  0.1892
#>  -0.3022 -3.8144 -3.4173  5.6919 -3.9242
#> 
#> (1,5,.,.) = 
#>    6.6940  -1.3223  -5.1221  -1.2939   4.8656
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]