Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   1.0080 -1.3679  2.5232  3.4663 -3.7060
#>   1.5225 -1.4674  1.7627 -1.1246 -1.8991
#>  -3.8224  2.4343 -1.0274  7.6483  0.1309
#>  -0.7291  2.5820  2.8007 -9.5619  1.5962
#>  -2.7396 -3.2944 -5.9192  5.7593  1.6888
#> 
#> (1,2,.,.) = 
#>   0.9945 -1.8653 -3.4946 -2.4970 -5.5332
#>  -2.2795 -2.5106 -1.4400  4.9874  3.3034
#>   2.9545  4.8488 -1.9399 -1.0220 -6.5010
#>  -0.4784 -1.0569  7.1008  1.4495  0.5088
#>  -3.9300 -0.6267 -8.2280  3.6718 -0.3948
#> 
#> (1,3,.,.) = 
#>   -3.4449  -2.2197  -6.3565  -2.8169   4.3917
#>    4.8901   3.5981   5.0130  -0.8816   0.8505
#>   -6.0137 -11.4551  -5.7762  -4.1098   8.6096
#>    0.5164  12.2493  -1.2664   1.1317  -6.5901
#>   -8.1494  -6.3558   5.6965  -2.2157   2.6218
#> 
#> (1,4,.,.) = 
#>    0.8238  -2.3813   1.0069   5.7276   2.6037
#>    0.6824   6.5187   2.6732  -4.3203 -10.5509
#>   -8.1670  12.8088   8.9468   0.4706   0.3906
#>    3.5992  -8.9463  -2.5184   1.2336   2.1534
#>    5.1990   8.6362  -0.8218   3.5677   1.0909
#> 
#> (1,5,.,.) = 
#>   -5.0145   3.4227  -1.0949 -17.3522  -7.9457
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]