Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   2.2046  4.0807  7.5487 -3.6397 -5.9399
#>   4.6385 -3.4319  9.7437  4.5418 -5.3972
#>   2.3035  6.8807  8.0219  2.5518 -3.9894
#>   4.2501 -1.2796 -2.8560 -3.1044 -6.8846
#>  -4.0370 -3.0117 -3.8804 -5.5187  2.9515
#> 
#> (1,2,.,.) = 
#>   0.7353  0.4296  2.7214 -5.1184 -1.2895
#>  -0.5150 -1.5934 -4.7894  3.8733  2.1250
#>   4.2016  3.5267 -0.3681 -1.3406  1.9816
#>   2.3497  6.9366  9.5330  2.5641  4.4814
#>  -4.5440  1.1848 -0.0263  8.3143  3.7930
#> 
#> (1,3,.,.) = 
#>   4.8957  4.6095 -2.5905 -3.7400 -5.0262
#>   7.1798  1.1212 -9.9754 -0.4614 -8.5377
#>  -0.4820  1.3440  5.3161 -0.1052 -0.1677
#>  -12.5107 -3.9337 -2.6621 -7.4047  1.0571
#>  -6.9298  1.0794  4.5538  8.1361 -0.4244
#> 
#> (1,4,.,.) = 
#>   -1.1230   2.0551   1.7777   2.7536   0.4799
#>    9.2050  11.6689   4.8528  -0.3193   3.8548
#>    3.1509  -1.4600  -0.2693  -5.6043  -5.3360
#>    8.1150  -0.4623   4.0847  -3.0933   1.6018
#>    4.0048   0.2208   6.4596  -0.3535   2.8429
#> 
#> (1,5,.,.) = 
#>   1.2007  7.5791  2.8725  7.4606 -2.5043
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]