Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   -4.0353  12.8197   9.2154  -5.4107   2.6977
#>    1.0771  -3.5323   2.6628  -8.0926  -0.2437
#>   -1.6471  -6.5611  11.5065  -4.1055   2.4605
#>    7.8753  -5.6637 -10.2128  -3.0754   1.1453
#>   -1.0517  -4.3298   7.8187  -2.1213  -5.2654
#> 
#> (1,2,.,.) = 
#>   0.9032  0.3430 -2.7233 -6.8786 -1.0918
#>  -2.1818  0.2474  3.1292 -0.7079  6.1573
#>   4.8648 -3.6855 -2.8263  0.3473  2.9087
#>  -1.1509  5.7876  7.0223  1.6745  4.6463
#>  -1.7867 -0.7749  2.2516  1.5591 -0.7145
#> 
#> (1,3,.,.) = 
#>   -0.2938  -6.3103  -5.0870   5.3048   2.3074
#>    3.1177  10.6432  -4.0184  -9.8932   0.0363
#>    5.3732   3.6696   3.7950   5.4341  -0.7093
#>   -3.3672  -4.8358   5.6849  -8.9064   0.4885
#>   -1.1463   3.9042  -0.4254  -1.1612   4.8481
#> 
#> (1,4,.,.) = 
#>  -7.5159  3.9637  4.5130  1.9114 -5.1685
#>  -6.7009 -6.7155 -5.6094  1.6618 -1.9255
#>   6.1821  0.5291 -4.9729  7.9547 -4.4640
#>  -3.6288  4.1249 -4.6011  1.8907  4.6955
#>   4.3717 -0.1645 -0.3984  8.5302 -2.0861
#> 
#> (1,5,.,.) = 
#>    4.1851   6.7425  -5.8291  -6.4638  -2.5490
#>   -7.7444   3.3598  11.2656   3.7632   4.9260
#>    2.0680  -8.3715  -4.6497  -1.5682  -3.3881
#>   -7.4167   9.3834   8.4867  -8.8878   0.4049
#>    1.7334  -2.0119  -0.4645  -2.7208  -3.3364
#> 
#> (1,6,.,.) = 
#>    1.8147   0.2100   1.0586   1.4795   0.3530
#>   10.1188  -4.6870  -6.3952   2.7602   0.6796
#>  -10.7326   6.0390   0.9987 -15.4185   1.7209
#>    7.0695  -8.3732  -4.0262   7.7217   3.8185
#>   -2.6632  -3.5939   0.3569  -7.0232   3.6756
#> 
#> (1,7,.,.) = 
#>   -0.2544  10.1754  -2.3259  -3.0603  -7.4932
#>   -8.2445  -9.2004  10.9605   9.6486  -3.0539
#>    7.9951  -1.3618  -6.2924  -6.6542  -3.3239
#>   -1.8144  -5.9885   8.7827   2.4808  -2.2207
#>    1.1304  -0.8414 -11.3034  -3.2052   2.5640
#> 
#> (1,8,.,.) = 
#>   -3.3671  -9.2297   2.7985   6.1013   1.7486
#>    2.3221   0.0839  -5.4045  -8.8113   1.0077
#>   -2.0208  -3.0169   5.3233   3.0406   2.8204
#>   -0.4338  13.2053  -6.7786  -8.4124  10.7062
#>   -0.9880  -6.7927   3.0104   3.1819  -1.0615
#> [ CPUFloatType{1,8,5,5} ]