Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -3.2380 -7.5985 -0.0150  2.9499  2.3165
#>   5.8705 -0.0581  5.8181 -3.3525  0.1480
#>   3.3454  3.9438  7.5383  0.5746  5.9159
#>   8.9169  8.5204 -0.6020 -0.8942  1.7330
#>   2.7585 -3.8858 -1.4645  6.2730  0.6392
#> 
#> (1,2,.,.) = 
#>   -1.0113  -5.4916  -3.4833  -3.2904   3.5458
#>    9.0589   5.4942   3.2675  -3.3090   1.5266
#>   10.7942   2.1215  -1.8738 -10.8653   3.7481
#>   -2.0272  -1.8975  -8.2387  -0.3907  -0.8264
#>    2.4988   0.4863  -5.2670   5.3227  -1.8820
#> 
#> (1,3,.,.) = 
#>   -1.8620   8.3953  -6.6764   5.9685  -0.0996
#>    6.7943  -6.3936  -7.2497  12.0893  -1.2527
#>    1.7360  -1.1437  -5.5699  -3.0688  -0.8162
#>   -0.3117  -3.9084   0.0779  -6.5031  -6.3492
#>    0.9984  -2.4064   4.5702  -2.1061  -1.1553
#> 
#> (1,4,.,.) = 
#>   7.9463  4.7114  1.8967  5.7438 -0.7535
#>   3.1639 -6.4880 -1.0632  0.9160 -2.8610
#>  -6.3206 -7.5533 -7.1201 -0.6040 -2.9129
#>  -5.8042  0.2040 -7.3014  3.0947  2.7872
#>  -0.9243  5.4747 -5.2613 -0.8317  0.8232
#> 
#> (1,5,.,.) = 
#>  -2.8530 -2.3514  3.4760  0.2204 -6.6908
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]