Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -2.8479 -3.9963 -4.9143 -3.6195 -11.5182
#>   6.6262 -3.0373  3.1239 -8.9967 -1.8047
#>   2.7999  1.3383 -10.7870 -8.5795  6.0278
#>   1.4088 -4.2918  1.6565  0.9413  7.7845
#>  -1.8269  0.2559 -2.1922  1.2250 -4.7587
#> 
#> (1,2,.,.) = 
#>   2.4364  2.0312  3.5766  1.8868  4.8140
#>  -7.0076  4.7970  3.8479  7.9306 -1.8578
#>   5.2262  4.2066  2.1968  5.5052 -8.6994
#>  -1.4239  0.0679  1.2653 -1.1399 -2.5896
#>   3.0515  1.3895  3.3529 -1.6711  6.5552
#> 
#> (1,3,.,.) = 
#>  -2.6469 -0.3541 -1.5763 -6.6843  2.4583
#>   3.8862  8.4937  2.9105  2.8182  3.2389
#>   2.7649 -1.8202  2.1360 -5.3240 -5.8861
#>   2.9728 -11.8622 -5.8040 -1.7442  1.2084
#>   3.1879 -3.4041 -5.5431 -0.2148  3.2997
#> 
#> (1,4,.,.) = 
#>  -1.7646  1.8303  1.6933  4.6008  0.8344
#>   3.0251 -0.6664  6.3955 -6.2255  6.7607
#>  -0.0451  0.3673 -1.3110 -2.3197  3.8975
#>  -4.0713 -2.9965  0.7477  0.8493 -1.3727
#>  -2.3044  5.2258 -6.6925  7.6762 -0.0283
#> 
#> (1,5,.,.) = 
#>  -3.4291  2.9156 -1.9658  2.5715 -2.0492
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]