Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -3.4092  -6.2466  16.4698   0.6137   1.0674
#>   -0.7827  -4.8015  -2.7845   8.6349  -0.2502
#>   -1.8986  -3.0787   7.6757  -6.7585   1.3876
#>   -4.6183  -0.8880  -1.2145   7.7465  -1.4755
#>   -4.0487  -2.1184  -4.0920  15.3129   6.0913
#> 
#> (1,2,.,.) = 
#>  -3.9970 -1.5973  6.3802 -8.4434  2.8217
#>  -2.4609  3.6255 -2.2682  3.9812  0.5355
#>  -2.9063 -3.0720  2.9116 -7.4635  0.7227
#>  -5.7243  6.2359 -11.2627 -3.1247 -4.9912
#>   2.5373  2.7397  0.8447 -0.3534  2.4941
#> 
#> (1,3,.,.) = 
#>    2.7612  12.3389  -5.9561  -5.1810  -7.3588
#>    0.7911   4.3272  12.5255  -3.0702  -5.8686
#>    0.6805  -0.4573   0.3831   3.4166   0.9818
#>    3.8068   1.9502  -7.4824  -6.2068  -7.7897
#>    4.8481   5.1738   0.9894  -3.6888   1.0473
#> 
#> (1,4,.,.) = 
#>    0.0962  -6.2705   0.9030   1.0252  -1.9850
#>    4.8836  -4.9744   0.5905  -5.4602   5.1137
#>    0.0349   1.8903  -8.8439  -3.0003  11.1338
#>   -0.1328   1.0919   5.1262  -6.1423   0.5675
#>   -2.5317   2.0305  -1.9586   0.3495  -8.9297
#> 
#> (1,5,.,.) = 
#>    0.8629   3.8769   0.8092 -10.1076   0.5506
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]