Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -1.0082   2.9970   2.4620  -2.5828   3.3189
#>    4.5294  -5.0766  -7.5891   7.0650 -12.1392
#>   -1.1922  -7.4472  10.2071  -6.3556  -3.4496
#>   -4.6205   3.6675  -0.8841   1.1367  -3.4481
#>   -4.7529  -7.6291   7.1899  -2.3027   0.0123
#> 
#> (1,2,.,.) = 
#>   -5.7239 -11.4803  10.2840   5.7233  -4.7855
#>   -0.7387  13.5937   4.4679 -11.9973  -6.0993
#>    7.5774  -1.0020  -1.8844   0.5265   6.6045
#>   -5.2048  -2.7610   6.9408   9.6409  10.1767
#>    4.9647  -2.9203   3.2924   5.9131  -0.1809
#> 
#> (1,3,.,.) = 
#>   0.4853  3.1435 -8.1767 -1.0026  4.9854
#>   5.8511 -1.9563 -8.5552  0.0762  0.0649
#>   1.5315 -1.5763  0.1705 -3.0600 -1.5757
#>  -1.0967  5.2266 -3.9278  2.1142  0.0746
#>   7.6189 -4.0100 -4.4663  4.2288  0.5377
#> 
#> (1,4,.,.) = 
#>   10.2372  -2.8354  -2.9636   1.8651  -4.0711
#>   -2.4473   2.8841  -9.6298   5.8257  -2.5337
#>    2.7311  -0.6682   1.4922   0.7956  -1.2774
#>    4.0060  -1.2521  -8.8803  -2.6058  -6.7359
#>    3.5167 -12.2235  -7.3769   5.3201  -1.0348
#> 
#> (1,5,.,.) = 
#>   -3.5709   3.7981  -1.0182  10.1998  -2.3658
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]