Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -1.4245 -12.7542  2.2528 -0.3866  2.5493
#>   2.9826 -9.0592 -5.1611 -10.6623  2.4037
#>  -1.4955  2.5807 -1.9657 -3.9358 -4.4436
#>   0.6805  9.7785 -3.2673  4.0095  4.2091
#>   0.9847 -0.4036  5.9882  9.3795 -1.7761
#> 
#> (1,2,.,.) = 
#>    4.5145  -2.4805  -5.3284   0.8233  -6.8563
#>   -3.3530   9.9481 -13.1431  15.9801  -6.0858
#>   -4.9445  -0.8279  -8.0712   5.7594   5.3230
#>    1.0521   5.7345   6.3723  -1.1788   6.5458
#>   -0.8010  -1.7652   4.9594  -0.4293  12.3910
#> 
#> (1,3,.,.) = 
#>   1.0681 -6.0911 -0.8935 -9.1405  5.0011
#>  -8.4795 -5.8684  7.3377 -5.9550  3.0338
#>  -2.6467  2.3587  7.7660 -9.2336  0.2182
#>  -1.8283 -0.6086 -3.1678  3.0317 -4.8309
#>  -4.2962  8.5252 -0.5398 -1.8218  1.2039
#> 
#> (1,4,.,.) = 
#>   -1.3445  -0.3195 -12.8374  -1.1816  -0.5928
#>   -3.2313   1.7963  -2.3132  11.4452   1.2307
#>   -7.9412   9.7058  -0.4148  14.5645   8.4712
#>   -1.6918   5.6646   8.0784  -0.2112   5.9365
#>    6.7174  -3.6893  -4.2169  -6.5144   2.1925
#> 
#> (1,5,.,.) = 
#>   0.2206 -2.0267  1.5345 -1.5597  0.3543
#>   2.1080 -2.1205  2.0761 -0.6247 -1.4351
#>  -1.2843 -2.1713  0.8377  9.0441  7.5876
#>  -2.2599 -8.1282  2.3085 -3.2934  0.4643
#>  -0.7322 -2.1198  1.5308 -5.6029 -0.2443
#> 
#> (1,6,.,.) = 
#>    2.5436  -5.5757  -5.1579  -1.0070   2.7929
#>   -1.8420   0.6654  -3.2031   3.6631  -1.7629
#>   -6.9052   7.3092 -12.1616  -1.1287  -4.9790
#>    3.0777  -2.6246  -4.7174  -6.3964   1.4545
#>    0.1741  -1.3698  11.8108  -0.2743   0.3037
#> 
#> (1,7,.,.) = 
#>  -1.8987  2.7313  1.7351  4.1529 -2.8448
#>  -0.9446 -5.0874 -6.1593 -1.2105 -5.3511
#>   1.0798 -2.2736  1.0058  5.0419  0.3171
#>   0.1284 -7.0855  4.5112  0.5907  5.5060
#>  -2.5361  3.2895 -3.2044 -2.4517  3.0299
#> 
#> (1,8,.,.) = 
#>  -1.2208 -7.6550  0.0724  3.7119  1.7695
#>  -3.7807 -1.3733 -2.9395  3.7062 -1.3440
#>   4.4652  5.5200  0.1732  4.1136  5.7837
#>   1.0496 -4.4151  3.9518 -4.2376 -5.2899
#>  -3.7516  0.0786 -3.3617 -11.5217  2.9681
#> [ CPUFloatType{1,8,5,5} ]