Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    5.9273   3.6302  -6.6493   3.5775  -1.7428
#>   12.0643  -3.0422   6.3004   4.2329   5.3249
#>   -2.1349  -2.9260  -2.1825   3.2933  -2.1370
#>   11.2849  -7.3356  -3.0232  -9.8265   6.6021
#>   -6.1352   4.1274 -10.2959   0.9415  -1.7891
#> 
#> (1,2,.,.) = 
#>   -6.3933  -4.7754  -7.3742  -1.8180   3.0997
#>    1.0066  -7.3329   6.7376  -1.9929  -5.1610
#>   -5.2756  -7.1600   7.5263  13.7584   4.2298
#>    9.9202  -7.7320   5.4639  -1.7522   7.2771
#>    5.4707  -0.4047   4.7542   3.1600  -0.2326
#> 
#> (1,3,.,.) = 
#>    5.9463   2.6235  -1.6817   6.4178  -1.0921
#>    6.1699  -5.2443  11.5710   4.9981  -5.5173
#>   -7.5775  -5.0030   0.2300  -6.2699   1.3399
#>    2.1591  -2.6726  -4.2979  -8.8437  -4.7063
#>   -2.7072   3.4339  -5.2751   6.3053  -6.3326
#> 
#> (1,4,.,.) = 
#>   0.7618  0.3098 -3.1035 -4.7677 -0.6906
#>  -6.1141  6.8376  9.3046  4.7053  1.2534
#>   0.8226 -2.2631 -0.2309  3.4252  0.2943
#>   1.5176 -2.1751  1.8335  0.4234 -7.0717
#>  -1.7997 -12.3120  7.5488 -7.2216  3.0622
#> 
#> (1,5,.,.) = 
#>    1.4548   2.8482   1.9259  -6.6422  -3.6829
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]