Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   2.0323 -2.7185 -4.7246 -2.9143 -6.1914
#>  -7.1922  3.9041 -9.4346 -10.0858 -1.0540
#>  -6.6939  0.8044  0.2491 -7.1196 -0.2504
#>  -0.3208 -8.1685  5.7261 -9.4578 -1.3223
#>  -6.1665  1.0945 -8.0199  0.9989 -6.3435
#> 
#> (1,2,.,.) = 
#>    1.0450   4.2699   3.3106   4.0855   0.5162
#>    1.9742  -3.5754   4.1204  11.5944  -1.5481
#>    0.5632   2.4178  14.0698   4.3916  -0.2552
#>   -4.6622  -6.4071   9.1215   6.3534  -7.1017
#>   -0.1992  -4.7133   3.7468  -2.0616  -4.3969
#> 
#> (1,3,.,.) = 
#>    8.0762   7.1825  -0.9253   3.6866  -0.1240
#>    1.7743   1.2424   4.8508  -6.2496   6.3861
#>    1.8372  12.8597   0.7701  -2.7742  -3.3992
#>    4.1329  -2.0625   5.9966  -2.3118  -1.3660
#>   -1.4455   2.7750   2.0484  -0.8667  -2.3288
#> 
#> (1,4,.,.) = 
#>  -2.8067 -8.5773  3.8130 -2.0588  1.6496
#>   3.2520  5.8688 -4.5708  8.4311 -0.2243
#>  -0.7814  2.9743  2.3795  2.2903  3.5286
#>   5.3478  7.4478  0.1746 -5.2242 -9.1970
#>  -0.6038 -1.3914  2.5480  5.2607  2.4263
#> 
#> (1,5,.,.) = 
#>   -8.0189   6.5668   0.4668  -1.0983   8.6738
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]