Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   0.1698 -2.5436  0.0947  4.0529  1.6757
#>   3.1217  9.4464  4.3920 -6.7086  8.4959
#>   1.0376 -4.1696 -7.6710 -4.8825  5.0163
#>  -0.1653 -1.6857 -8.0928 -3.7481  8.8435
#>  -7.1461  6.9997 -3.9453 -0.3677  9.8846
#> 
#> (1,2,.,.) = 
#>   -9.4971  -4.9520   4.0231   6.2464  -0.9963
#>    0.0139   2.8952   2.9869  15.2876   2.9009
#>   -4.7154   1.0337  10.8808   5.1344   0.4212
#>   -4.3152  -4.6552  -6.4583  -2.6049  -7.4229
#>   -6.4623  -3.5868 -11.1954  -5.8146  -3.4394
#> 
#> (1,3,.,.) = 
#>  -1.2106 -7.0214 -1.0720  1.2772 -4.4254
#>   0.4767 -2.3032 -2.5804 -0.2402  1.7999
#>   5.7964 -7.1256  6.3867  8.1562  2.1218
#>   4.0069  0.2775  0.5378 -3.9324  5.8137
#>   2.7873 -6.7611  4.4332  2.0886  7.5268
#> 
#> (1,4,.,.) = 
#>    0.1832  -2.0472   1.9408   4.3962   1.6350
#>   -1.2650   3.2793   1.6304  -0.5583  -3.4907
#>    0.2604  -5.0199  -5.6403  -5.5993 -14.0558
#>   -5.5548  -8.9967  -4.3301   0.2537  -7.8864
#>   10.0234  10.6502   2.2274   9.7194   0.7478
#> 
#> (1,5,.,.) = 
#>   -5.6892   3.1916   0.5291  -0.5226   2.1267
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]