Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   3.9759  4.5270  0.3876 -1.2829  1.2035
#>   1.0273 -7.7417 -10.0587  1.6671  3.0458
#>  -4.1567 -7.1540 -6.8642  3.7737  4.5401
#>   5.2519  6.5644 -7.9045 -2.1787 -0.2772
#>   7.0424 -3.6253 -5.2846 -8.9427 -2.3551
#> 
#> (1,2,.,.) = 
#>  -0.0268  0.6754 -4.6242  3.2899 -2.3975
#>  -0.9419 -1.1215 -1.8557 -1.5100  4.3690
#>   5.7359 -5.4696  3.5256 -2.9685 -6.9045
#>  -5.0968  5.6396 -2.9441 -1.3425 -1.0743
#>   8.6178 -1.6334 -1.4886  0.6269  4.1990
#> 
#> (1,3,.,.) = 
#>    1.3357  -2.0058   1.7542   2.9037  12.0313
#>   -7.4289  -0.9628  -6.6927  13.0895   7.2721
#>    5.8602  -1.8625  -8.7953 -11.6021  -1.2041
#>   -0.4511   3.7897   5.4251   2.9262  -5.9291
#>   -2.3349  -3.9883   1.9682   5.2461  -0.8450
#> 
#> (1,4,.,.) = 
#>  -1.2250  3.7687 -0.1534 -3.4252 -7.6688
#>   1.8466  5.3550  8.4317  6.0137  1.2339
#>   1.2167  1.4621  5.0784 -5.5934  2.6860
#>  -2.0950  3.0802  4.6563  1.5367  5.0845
#>  -3.1924  0.5774  1.7143  4.7431  1.6757
#> 
#> (1,5,.,.) = 
#>   -0.5078  -6.9827  -6.7323   0.4076   0.7693
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]