Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -5.1921  5.5357  5.3286 -2.0196  2.7888
#>  -2.6769  2.9594  4.5634 -3.0055 -3.2475
#>  -1.7778 -5.2756  3.1237 -4.0470 -2.2169
#>  -9.1741 -10.2130 -2.0137 -3.5302 -2.7381
#>  -1.1242 -1.4073  1.4610  2.6191  3.4224
#> 
#> (1,2,.,.) = 
#>   8.3695  3.7232  1.2493  1.4385 -2.9323
#>   7.0157  1.9971 -1.6405 -1.6115 -2.4082
#>  -2.7022 -3.4308  6.9191  3.2954 -2.1334
#>   2.3466 -7.2441 -2.2841 -6.1483  0.1179
#>   8.5449 -1.9946 -3.9945  5.9688  2.9611
#> 
#> (1,3,.,.) = 
#>    3.1762   4.1004  14.1245  -3.6587   0.7456
#>   -4.3642  -3.4130   1.1430  -3.6927   0.3611
#>    2.7103  -2.1818  -9.0248   2.6369  -6.7434
#>   -7.5113   4.6478  -6.2653   6.3886   0.9126
#>    0.5359   3.9746 -10.5266   0.2615   5.8921
#> 
#> (1,4,.,.) = 
#>  -3.9083 -1.1065 -1.0743  8.8988  1.5515
#>   1.4525 -6.7584  5.6658  4.9486  2.2946
#>  -2.3814  0.3884 -11.8000 -7.6908  1.5716
#>   0.6945 -11.1542  9.4895 -4.6394 -0.7710
#>  -6.1592  3.9215  6.8214 -10.0025  3.6498
#> 
#> (1,5,.,.) = 
#>    3.8952  -5.5058   4.2107  11.3144   7.5824
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]