Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -1.2197  0.1042  4.1963 -5.7578 -1.9155
#>  -11.1401  3.9720 -1.1968 -3.3774  3.5295
#>  -3.5278  3.9326  1.6033 -1.0749 -0.2698
#>   3.6050  6.5481 -2.0061 -3.5733 -6.1433
#>   2.4057  6.2438 -3.0705  4.1704  3.6954
#> 
#> (1,2,.,.) = 
#>   0.0657  6.9108 -3.2564  0.9620  0.9335
#>   0.9287 -8.2590  8.3011 -8.3294  1.8537
#>   0.9816 -1.6492 -6.5055  9.0605 -3.1920
#>  -4.7447  2.0515  6.9643  3.3202  1.7706
#>   2.5334  6.7776 -3.4976  2.4046 -3.6640
#> 
#> (1,3,.,.) = 
#>    8.1730  -7.9311  11.6332   0.4559  -3.0773
#>    3.6482  -6.2172  -6.2961   5.6774  -7.6801
#>   -7.6839  -4.8997   6.0907  -6.0771  -0.2029
#>    3.2188   5.2025  -1.0159  10.1948  -0.5114
#>   -1.1870   2.4088   4.8312  -6.1595  -8.6275
#> 
#> (1,4,.,.) = 
#>   -1.7704   2.8928   3.1927   1.0221   0.0123
#>   -6.2117   2.1562   3.8502  -3.0905   6.2229
#>   -2.9163   3.7950   0.0478   3.6391   3.2088
#>    1.0252  -4.1659  -2.1801  -4.6814   3.6963
#>   -0.6667  -3.6708   0.1713  10.9406   7.3032
#> 
#> (1,5,.,.) = 
#>   -0.6735   8.3934  -5.0698   1.0284  -0.6702
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]