Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -0.6896 -1.0543  4.6986  1.8924 -1.8985
#>   2.7988 -6.0373 -8.4634  2.8179  3.5334
#>  -5.0004  6.0514 -4.4466 -8.2887 -5.8974
#>   8.3628  4.1695 -8.7238  5.1077 -9.4228
#>  -4.2653  6.2148 -1.3670 -6.3311  2.0987
#> 
#> (1,2,.,.) = 
#>   -6.3910  -3.6027   5.0474   0.0040  -7.5475
#>   -0.3071  12.7042  -0.1289  -3.7233  -3.8144
#>   -2.6420   1.6090   7.4236 -11.8012   6.9226
#>    1.1131   4.7250   4.1696   9.1579  -4.1600
#>   -2.5305   0.4865  -4.5986   6.3792   2.7622
#> 
#> (1,3,.,.) = 
#>  -2.5574 -5.0432 -1.1274  0.8890  2.9802
#>   7.2206 -6.1683  4.7874  9.6772  1.2632
#>   8.0079 -7.8945  1.9981  6.9076  5.4451
#>   3.7574  2.2239  6.9871 -3.6611  9.3037
#>   0.0821 -0.0513  5.1705  2.9172  8.0703
#> 
#> (1,4,.,.) = 
#>   -6.5426  -0.7670  12.3715   5.6016  -1.6671
#>   -0.2765   0.0516  -7.4371   2.7947   1.7979
#>    4.5755  -6.5390   6.4556  -6.7493   0.2759
#>    7.6899  -5.4547  -4.4699   8.4932   6.7214
#>    3.1438  -1.8379  18.2276  -7.2014  -5.6200
#> 
#> (1,5,.,.) = 
#>    1.4059   7.8880   6.5247   0.7758  -1.0018
#>   -2.3757  -7.0608   0.7943   7.3854   2.5127
#>    2.7627  -5.5286   0.3717  -5.6946   4.3778
#>    1.3917  -5.0810   7.5181   6.1666  -3.1988
#>    2.0664  11.2272  -6.7472   2.6550  12.0951
#> 
#> (1,6,.,.) = 
#>    3.8145  -1.1549   2.8868   5.2840  -4.6121
#>   -4.0756   9.7043   5.4599  -6.6687   4.1325
#>    0.8876  -2.1477  -4.5167  10.0116  -8.6302
#>   -1.5642   5.1446  -6.3055   3.7346   1.2333
#>    9.7478 -10.0418   1.6803  -0.1517  -8.0267
#> 
#> (1,7,.,.) = 
#>   -2.0394   1.7432   2.2109  -8.0309   0.4374
#>   -4.0676  -5.8783   0.8601   6.0830  -5.4613
#>   -4.6126   3.5520   3.7185 -13.5229   2.4259
#>    1.5730  -1.8012   0.4879  -2.7340   3.9698
#>    4.7576  -6.3271  -0.9755  11.2212  -5.1217
#> 
#> (1,8,.,.) = 
#>    4.0968  -8.1191   7.5766   2.4911  -0.4617
#>   -7.3530   4.9422  -2.8665   0.4612   2.0061
#>    9.6341 -20.8767  18.0032 -12.0646  -3.8855
#>    6.9379  -7.5581  -1.9725   2.6473  -6.0724
#>    8.5012 -11.4259   6.6720   2.6199  -5.9310
#> [ CPUFloatType{1,8,5,5} ]