Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    4.0963   0.5904   1.3010   3.3888   3.7700
#>   -7.2961 -11.0085  -9.8475  -0.7299  -2.9434
#>    8.8163   6.7214  -7.2088   4.5503  -5.4505
#>    0.4084  -3.9398  -3.3048  -8.0228  -2.0116
#>    0.9227   7.1012   2.0623   5.0928  10.7367
#> 
#> (1,2,.,.) = 
#>   0.1024  0.2484  0.0975  2.5095 -1.8186
#>  -2.6639 -2.4149 -1.3023  2.2383  1.5583
#>  -1.7805  3.3121 -4.8707  7.9709  4.6873
#>  -0.7015 -1.6208 -6.0083  4.0122  5.6453
#>   3.4666 -1.0481  2.5026  1.7927 -2.6220
#> 
#> (1,3,.,.) = 
#>   -0.0473  -0.6522  -3.6196   0.7729   2.2327
#>    0.3747  13.2102  -3.2206  11.3733  -2.3560
#>    1.7056   2.2142  -0.6974  -0.0017  -2.4776
#>   -3.4955   0.5256   5.6663   1.0480   2.4002
#>    1.1099  -0.2473   0.7442  -2.8684  -2.2373
#> 
#> (1,4,.,.) = 
#>  -7.9109 -12.2198 -6.5733  5.6032 -7.4729
#>   2.1952 -8.8235 -6.1461 -9.7207  3.7734
#>   5.9111  7.3537 -1.9087  0.0209  6.1159
#>   6.0473 -0.8405 -3.8195  0.4570  4.0590
#>   5.3143  4.4756 -0.0272  8.5616  2.6294
#> 
#> (1,5,.,.) = 
#>    5.3963  -8.6214  -4.3421   6.2431   3.9505
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]