Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   6.6923 -2.6407 -8.3372 -2.7472  0.3312
#>   5.6129 -0.7655 -18.3635  6.7531  5.9059
#>   0.2328 -3.3934  2.0331  4.7988  7.5138
#>  -1.2359 -0.1631  7.5524 -4.2232  6.3918
#>  -3.8908  0.1743 -1.1712 -3.0208  6.5735
#> 
#> (1,2,.,.) = 
#>   3.7232 -4.8422 -4.5066  8.9546 -2.6469
#>   1.2111 -1.8118 -0.4600  9.0296 -6.1581
#>   1.5054 -5.8997 -1.5585  4.5409  6.0696
#>   0.2827  7.2237 -3.5909  8.5655 -4.3517
#>   2.1768 -2.6398 -3.6925  9.3171 -5.1362
#> 
#> (1,3,.,.) = 
#>    1.5062   1.8435  -0.7324  -2.1743  -2.8973
#>   -0.0804   2.0236   5.2378  10.1878  -5.5977
#>    1.4759   4.6194   1.1066   2.3725  -5.7608
#>    8.6985  16.0952  -6.9954  -4.5836  -9.8161
#>   -5.8418   9.0493   8.9210   2.6071  -7.4965
#> 
#> (1,4,.,.) = 
#>   7.4119e-01 -3.3681e+00 -5.5234e+00 -9.6681e+00  2.3512e+00
#>   5.8778e+00  8.8462e-01 -3.7941e+00 -8.2050e+00 -5.0104e-04
#>   4.3675e+00 -2.8480e+00  1.0759e+01 -1.3041e+01 -9.0878e+00
#>  -1.4124e+00 -1.7373e+00  2.1847e+00 -3.6982e+00  3.2902e-01
#>  -1.5190e+00  3.6488e+00 -1.1560e+01 -6.0818e+00  7.1263e+00
#> 
#> (1,5,.,.) = 
#>  -0.5836  2.0356 -2.3079 -0.8670 -3.1881
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]