Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -4.8317   0.0717   1.8242  -2.4354  -0.1416
#>   -6.6307   6.0931 -16.3305  -0.5607   9.5992
#>    2.2583  -7.8056  -3.0085  -2.3534  12.7793
#>   -7.3489   8.5292  -2.8568  -0.8678   6.9324
#>    4.1843   0.2911   6.2647 -10.0391   4.2844
#> 
#> (1,2,.,.) = 
#>  -4.2800  3.0278 -1.2474 -6.8382  1.4996
#>  -2.5907 -3.0229 -1.7026 -1.9611 -2.3048
#>  -3.0963  0.6615  0.8244 -2.0539  0.1646
#>   5.3175  6.6419  2.0163 -1.9063  4.5564
#>   0.8130  3.9282 -0.6597 -1.2209  4.7872
#> 
#> (1,3,.,.) = 
#>  -3.2137  4.9367 -2.2517  1.0498 -3.5973
#>  -0.0261  5.7577 -5.3192 -2.9249 -3.8128
#>  -4.5708 -3.6611 -2.3508  7.5701  6.6584
#>   3.4774  0.0338  6.3015 -2.2378  1.4739
#>   0.2060 -2.1721  0.2027 -8.6794  3.9166
#> 
#> (1,4,.,.) = 
#>   3.1915 -4.5808 -1.8154 -3.6778  2.5264
#>  -8.6356 -3.2008 -1.5383  2.5210  4.8553
#>   4.2540  7.9529 -0.4436  2.4200 -0.1807
#>   4.2097  4.8300 -3.5218 -3.1653  3.3154
#>  -3.3612 -13.2833  3.3967  5.6302  3.1397
#> 
#> (1,5,.,.) = 
#>   2.6524  4.3286 -3.3243  2.5129  0.3775
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]