Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -5.0283  0.1602 -2.8069 -6.9584  0.2346
#>   8.7712  8.8503  9.1932 -3.9359 -0.1210
#>   5.5094  3.0214  5.5884 -0.3030 -3.3806
#>  -3.6080 -4.5101  0.5240 -0.6497  6.0954
#>  -3.3437 -0.4410 -1.4863  0.4044  0.2735
#> 
#> (1,2,.,.) = 
#>    0.4639  -4.6963   8.0539   2.6611  -0.2630
#>   -2.6222  10.4801  -0.0708  -0.8211   6.5695
#>   -8.1312   2.6588  -6.6414   4.8359  -1.0666
#>    1.6027  -5.4744  -0.8058  -2.6885  -2.4875
#>    0.4074  -0.1539   2.4496   0.7759   0.9051
#> 
#> (1,3,.,.) = 
#>   -8.3058  -2.8982   3.9080  -1.1610  -2.5816
#>    0.2607  -2.0604  17.3121  -6.9569   4.7975
#>   -5.7037   0.2910  -0.2114   1.6064  -2.9894
#>    0.3293  -1.3569  12.6577  -2.4029  -4.8191
#>   -1.7121  -0.1121   5.9211   0.6520   3.5331
#> 
#> (1,4,.,.) = 
#>   -0.5677 -13.3817   5.4724  -8.6836   9.2921
#>   15.9482 -13.2471   0.6615  -4.0440  -4.7738
#>   -3.6449  -4.5335  -4.6355   3.0678  -6.9127
#>   -6.9050  -4.4610  16.4135   1.2793   7.0406
#>    0.9351   3.8226  -1.1875   4.4145  -5.2016
#> 
#> (1,5,.,.) = 
#>   -1.6923  -2.7517 -11.4179   1.7319  -3.2105
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]