Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   4.8102 -4.5484  7.1301 -2.0937  4.8638
#>  -8.1503  7.2391 -9.0001  9.7714 -8.3653
#>   1.0434  1.2375  2.3423 -8.6359  5.8947
#>  -0.8400  3.7177 -4.6217  1.7326 -1.9220
#>   0.0011 -0.1329  1.4485 -3.9339  2.8734
#> 
#> (1,2,.,.) = 
#>  -2.0796  4.4619 -1.4301 -5.5153 -5.6926
#>   6.3563  7.4598  2.6847  2.2523  5.7524
#>   4.5013 -1.4680 -5.3717 -4.1847  3.4696
#>  -3.6180 -9.6369  5.1639  6.9169  7.6861
#>   1.9839 -0.6636  3.4117 -2.5449  0.3833
#> 
#> (1,3,.,.) = 
#>  -10.2893  -0.8659 -11.7780   5.5874   3.5406
#>   -4.0088   5.7267  -3.9193   4.2709  -2.1667
#>   -7.2175 -11.4053  -0.1966   6.3381   6.7138
#>    1.3441  -2.0153  -6.3973   2.2502  -4.3356
#>   -2.6781   4.4470   2.7626   5.2420  -2.1647
#> 
#> (1,4,.,.) = 
#>   5.6309 -4.6866 -6.3087 -1.5008 -4.6611
#>   3.5336  2.0712  1.7756 -2.9609 -4.2367
#>   2.0818 -0.6041 -0.9293 -6.1727  3.5395
#>   7.4497  3.2332 -5.3575 -8.0566  1.1478
#>   1.4862 -2.1563 -0.7200 -5.0050  5.2937
#> 
#> (1,5,.,.) = 
#>  -13.2442   2.4257 -14.1366  11.0087  -2.4551
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]