Conv2d

Conv2d

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -2.8890  11.1146   1.1645   1.2773   0.4935
#>   -0.5150  -7.5357  -1.1350  -0.2700  -0.5519
#>   -3.0395   6.9926  10.5779  -1.8876  -5.2639
#>    3.8827   2.9881  -8.0971   6.4604   5.8950
#>    5.7192  -9.7082   1.4499  -0.0692  -1.6573
#> 
#> (1,2,.,.) = 
#>  -0.2325  6.6339 -0.7890 -4.7958 -1.8376
#>   5.4961 -2.9121 -7.5725  0.1189  3.1303
#>   2.6527  3.6286  8.3872 -0.3904 -4.2168
#>   0.0426  6.9334  0.1632  2.2246  6.9478
#>  -5.6071  1.5920  2.0265  3.6582 -2.9247
#> 
#> (1,3,.,.) = 
#>   -0.6336   5.0524  -3.0567  -1.6929   1.0053
#>  -10.0534  -7.2477  -3.9543  10.8954   4.2620
#>    5.4695   4.9607  -3.9879  -2.0448  -7.6528
#>    9.1341  -3.0277   0.6887  15.5823   1.9411
#>   12.1133  -6.9714   4.3026   2.5530   4.0191
#> 
#> (1,4,.,.) = 
#>   -8.2275  -2.9306   6.7626   1.1146   0.8171
#>    2.4885   7.6537   3.0743 -12.9992   0.4673
#>   -1.8984  -1.3299  -4.8658  15.3498  -0.2197
#>    6.1324  -5.1148  -5.8368  -8.6798  -6.6626
#>    0.5922  -1.1777   0.6139   3.0273   3.7171
#> 
#> (1,5,.,.) = 
#>   2.0613  2.5446  0.6547 -3.0989  0.0041
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]