Conv2d

Conv2d

Usage

torch_conv2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  dilation = 1L,
  groups = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{out\_channels} , \frac{\mbox{in\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias tensor of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

implicit paddings on both sides of the input. Can be a single number or a tuple (padH, padW). Default: 0

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

conv2d(input, weight, bias=NULL, stride=1, padding=0, dilation=1, groups=1) -> Tensor

Applies a 2D convolution over an input image composed of several input planes.

See nn_conv2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
nnf_conv2d(inputs, filters, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -3.4059 -2.7058  2.5093 -3.4216  2.9617
#>   2.5063 -0.7043 -4.4236 -1.0526  7.4263
#>  -7.5026  3.0381  3.7541 -2.4213 -3.8267
#>  -1.2342 -0.9014 -2.8035  2.2455 -1.3639
#>   2.4405 -0.0463  2.4081 -3.7785  3.7052
#> 
#> (1,2,.,.) = 
#>   -1.2265  -3.3740   6.2522  -4.7079  10.6204
#>    9.4865  -1.4172   4.3375  -4.4431   1.0612
#>   -3.7765  -1.9476   1.9869  -1.8271   1.5890
#>   -2.5165   2.6892  -3.0189  -1.2474   2.9263
#>   -0.8248   3.8496  -5.1305  10.1645   5.9079
#> 
#> (1,3,.,.) = 
#>    3.7720   2.2191   4.1607   4.5565   3.3336
#>   -4.1401   7.8400  -1.0436   1.4861  -0.2996
#>   10.3055   6.9497   1.8176   3.5512  -7.2518
#>   -0.4999   6.9724   3.3465 -10.3968  -2.1061
#>    2.9363  -5.7066  -1.7894  -4.3423  -0.9909
#> 
#> (1,4,.,.) = 
#>   0.6043 -1.4250  9.3753  3.3690 -4.3549
#>   7.8421 -4.2227 -6.3710  4.1784 -1.2735
#>  -0.6254  3.8438  9.5027 -9.5677 -12.3726
#>   3.1261  4.9193 -8.7281 -7.8797  3.9891
#>  -1.1519 -7.6296 -3.7725 -2.1500  2.5550
#> 
#> (1,5,.,.) = 
#>    2.7502  -0.5347  -3.5698   2.2831  -1.8657
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]