Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.4774   8.3218   1.4108  -8.2997  -7.4163  -3.9211  -0.8174  -6.8326
#>    0.4993  -2.0364  11.8241   8.1041 -13.8811 -10.2164  15.0798   6.9951
#>   -6.3050  -1.3402  -5.8989   7.8864  10.8328  14.5007 -12.2190  -4.2921
#>   -3.1156   2.4620   6.7459  16.3698  14.4141  -1.1328   4.2522  18.6630
#>  -10.7594   3.7227  -8.0192 -10.3243   2.0398  14.7523  16.1442   5.9109
#>    6.6723   2.6970  -6.8695   7.1520  -7.0714   9.4469  -7.4526  -2.8301
#>    0.0094  -0.7640  -7.6490   8.3591  -9.2201   4.1235  -1.5783   3.1686
#>    2.3387  -1.3956  -6.9814  10.2847   5.7309  11.3040  -2.2081  -8.8180
#>   -1.5215   5.7262  -4.4082   1.2191 -13.4829 -11.8141 -14.1962  -7.2076
#>   -0.6018   9.0084  13.1689   5.0397  -9.9269   5.2146  -5.8709 -18.8801
#>   -2.8096   0.0994   8.9398  15.4323  -4.1640   6.8414   7.0630  10.7604
#>    0.1206  -0.5341   0.6079 -16.6005 -22.5958  -2.5495  -5.3863   2.0979
#>   -1.5880  -5.8452  -4.9054  -6.2206 -26.1390   2.4894  -3.3319   4.6648
#>    0.9162  -9.2935  -8.5445  -3.7573 -10.2101  10.3845  13.6845   6.1481
#>    1.5113   5.4629  -3.7426   6.8143   0.2407  14.1018   4.0150   1.6099
#>   -0.7173  12.1432   1.2767 -12.8029  12.5760  -1.4143  -3.1479  -2.5847
#>    5.0082  -9.9579   2.6504   7.8569  10.2480   5.9441  -9.1595  -5.5374
#>    4.1052   0.0702  -2.4444  -0.2149   9.6187   8.6116   1.1701   0.3904
#>   -6.1685   6.4417 -22.4567 -18.3909  -7.3681   3.4497   8.4756  -0.7684
#>    6.3563   9.5405   8.4754  -1.1953  -1.2396  -1.1445  11.3529  12.0383
#>   -0.4950   9.3065   2.0458   7.2560   6.3952  -3.9229  11.0552   7.0072
#>   -6.5900 -12.4919  -0.4432 -10.4470   2.6776  12.1000 -17.8946  -9.8785
#>  -11.2404  -9.6185   0.4392   7.1898   9.7719  -0.7628  -1.9516   7.8014
#>    2.3451  13.9515  -4.7366   7.1424  -4.3046 -18.3213   7.7414  -7.4726
#>    1.9505   4.8621   3.2876   1.4858  -9.4166  -7.9584 -17.7835 -13.1978
#>   -4.2885   2.4203  -5.0089  -0.9957  -2.5203   1.3278 -12.7652  -3.4486
#>   -1.7265   2.4457   1.3962 -11.9053  -6.2060 -15.8478 -16.9617   2.9888
#>    5.5506  -7.7391  12.3395   0.0130  -4.8862  10.4603  -8.8612   5.0794
#>   -7.2915   1.6260  -9.6010  16.6345  -0.2907  -1.1868 -13.2613 -13.5281
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]