Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.8350   2.6149   2.0106   1.7460  -0.7047  12.1228  -6.9216   3.1252
#>    1.2273   0.6921  -0.9145  -0.5745   5.4487  -2.0442   3.1931  14.6537
#>   -1.3550  -5.8976  -1.8467   7.1141  -9.0305 -13.0667  10.8825   5.8220
#>    6.0031 -11.8067  -2.2186  -1.4089 -11.0698  -8.3585  -3.9299  -6.2901
#>   -3.9519   0.1547  -5.2165  -7.7373  -4.4767  17.7629  -9.6000  -2.1530
#>    3.5102  -3.5480 -13.7224   0.1011  -2.1462  13.4356   4.9314  -8.5296
#>   -0.5616   5.7595  -5.7855   4.3410   5.7034  -2.0412 -28.3252  11.5557
#>   -2.1385  -3.9466   4.3258   1.5312   0.4830  14.1378   0.3256  14.7396
#>    0.9334 -11.0584  10.2454   1.1441  -8.4763 -12.6573 -17.6403  -7.8599
#>    1.2496  -1.7424  -0.8985   2.2252 -12.0868 -12.8998  -5.9843  -1.4307
#>   -2.4819   6.0316   7.8697  -6.7486   3.6959 -19.5486   0.8588  13.1949
#>   -1.2537  -7.8784  -6.6862  -2.6250   7.7154  -4.8763   0.0400 -15.3147
#>   -4.5275   0.8246   8.4244   2.9673  -8.0905  -3.4521  22.7647  -0.4625
#>    4.3328   2.0250   1.8371   0.3538   7.2539   8.3323 -11.6136   9.3823
#>    0.8993  -6.5690  -7.2823   6.0717  25.4928  -0.0576 -12.6779  22.4511
#>   -0.4593  -3.9731 -10.1031   4.9817  -1.5248   6.5220  11.1697  -2.8754
#>    1.6746   4.5015  -2.7915  14.9360   2.4739  11.7647  12.6771  -1.6430
#>    0.7298   2.4126   1.2963  -3.1108   8.6550  11.4145  11.2673  16.5715
#>   -2.8255  -0.3262  -0.0978  -9.9089   2.6927  -3.9764 -19.9485  -5.3375
#>   -2.8915   5.7983  -1.4856  -7.7156  -0.0173   7.8446   0.9514   3.8097
#>    2.5616  -3.7595   6.3990   1.0906  -3.4577   3.4105  -8.9603  -7.8687
#>    1.7844  -0.6063  -2.2253  -1.9668   0.9936  13.7163  -1.8220  -9.4336
#>   -0.0573  -1.8051   4.4483  18.2280   0.8481   2.1699  -4.8503  14.0259
#>    0.7536   2.9141   7.4740  14.1385   6.0521 -11.7223  -2.8487  -3.5959
#>    2.1402   4.3631   4.8714  15.2373   2.2326 -13.6625   2.2344   1.2816
#>    1.6130  -4.0998  -5.3476  10.1978  -0.6463 -17.0464  11.9474  -8.5898
#>   -5.5678  -4.5184  -4.4866   0.9482  -3.1510  18.6158   7.1995  -9.9210
#>   -2.4828   1.4119 -22.0227 -14.8069  16.5673   6.2327  -3.5733  10.1250
#>   -0.2125   4.2134  -0.4397  -0.5984   9.8926  -1.0857   3.9126  -0.8802
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]