Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.2890  -2.6807   6.7701  -3.8718  -1.0493   2.5676  -2.9120  -4.6018
#>   -1.1814   0.6336  -1.2037  12.7938  -9.6454  -7.6724 -16.4952  -2.5464
#>    4.0883  -4.3313   3.6351   2.4271  -8.1361   1.6323  10.2652   3.9520
#>    5.0560  -5.3205 -12.8047  -8.1523  12.9062  -5.6710  -4.6998  -4.8460
#>    1.9949   3.3181   3.6198   7.6778  -9.6297  -0.5206  17.5107  22.2250
#>   -0.5646   1.7745   0.1454  11.2494  19.4388 -17.3421   2.4460  -3.1442
#>    6.7427  -4.7694   0.7100  -9.4870  13.1184  -7.8456  -0.1727 -12.4206
#>   11.4242 -10.1517  -1.5643  -3.9375   0.5415  -5.6855   5.1289  -4.1314
#>    0.3085  -1.1570   9.6745   7.9448  -3.4779  -2.5028   6.2180  -9.0859
#>   -2.1454   8.8400   0.5130   5.4233  -0.7159  12.0768   2.7442   1.2996
#>   11.1002  -5.4208  -5.9856  -7.8985  -3.5454   7.0592   4.1517   2.2176
#>   -5.9763   5.9577   5.7460  -3.9230   2.5662   3.7366   2.2437  12.0031
#>   -3.8046   1.3453  17.5494  -5.8863  -9.8391 -13.2734   1.4607 -11.0094
#>    1.4289  -4.6188   4.4997  -0.8285   6.4614   1.4358 -10.5386  -5.2524
#>    9.4552  -2.9788   7.1182  -9.8283  13.0622  -6.5079  -3.5063 -12.8781
#>   10.7776  -8.5911   5.5285   3.1138 -12.9805 -11.9142  13.0108  11.3299
#>   13.3208  -8.1458   9.7359  -2.3863   4.4158  -3.3582   0.8763  -7.7580
#>    3.3002 -10.3918  -2.9696  -0.7502  -9.9030  -4.7124  10.2624 -13.4670
#>   -3.5619  12.0292  -6.7829   8.5159  16.7139 -23.3040  -0.8500  -4.0140
#>   -1.6695 -11.9479   3.6300  -2.0283  12.2704  -7.5727   1.8449  -7.3874
#>   -3.2200  -1.2557   4.7662   4.7845  12.9935  16.1410  -0.3327 -18.9777
#>    8.2003  -8.6922  -0.1152  -4.2042  -1.3381  11.3541 -11.4744  -4.6434
#>   -3.6808   4.1723   1.0115   8.1602 -10.1970  -5.8974   1.0128   3.0476
#>    0.1729 -13.9779   7.8464   5.2125   6.7179  -5.2539 -14.4543  -4.1230
#>   -5.0996  -4.0936  -7.5974   1.6002   4.1222  -0.6183  -0.1356 -10.1587
#>    1.9463  11.7067   5.5819   3.7488   3.2211  -4.7771   9.7279   8.6904
#>   -4.9238   3.0903   6.0952   2.0259 -18.2732  -4.2571  -7.6368 -12.5821
#>    0.9046  -7.3512  -9.3602   5.7232  21.3271  -5.3640 -16.7231  -4.6245
#>   -0.9687  -7.9309  -0.8461  -4.0537  20.1685  -6.0676  -7.3074  -1.1030
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]