Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.7658   8.2431  -1.2703 -20.0981  -1.5739 -12.1844  -7.5455   3.4006
#>    3.8789   2.8798   0.6160   2.0801 -10.1592   2.2419  -6.9915  -2.2541
#>   -2.0159  -1.2353   7.1388   0.4792  -5.2595   6.8064   6.3957  -2.8658
#>    2.9314  -1.9784  -9.4239 -16.8028  -7.4604  -8.8974  -9.1202  -0.7140
#>    0.5037  -8.3349   8.1708  12.4064  -4.9011  10.7827  11.3293  11.0721
#>   -4.2615  -9.0657   5.3262  12.8149  -6.3501  -2.7691   7.2708   3.3713
#>   -3.7826  -1.4823   5.1202  -0.1718  11.4869 -11.1204  -7.8776   4.6669
#>   -0.3163   4.8865  10.0571  -2.8802  -5.3131  -1.5786  -6.4918  -4.2405
#>   -1.2680  -4.7212  -4.9792  -4.1067   3.1683  10.9806  -5.4988   7.5210
#>   -0.0691  -3.3428  -3.7314  -8.3910  -2.0025   0.4396   4.9085   4.4734
#>    4.6457   2.3872  -0.5930  -3.5125  -6.4509   8.7602  -4.7180 -11.6610
#>    0.5096  -4.1882   0.6815 -10.5104   1.7592  -1.3172 -13.8128   0.6509
#>   -0.3423  -5.0948  -2.7439  -0.9475  -7.1491  -2.4461 -15.1090   4.7021
#>   -1.3963   0.5510  -2.3494   3.9969 -16.3294  -0.8168   2.5181  -8.1803
#>   -3.5452  -0.6335   1.4526   9.4931 -11.2285  -0.8240  -6.1238  12.1671
#>   -3.2606   6.3002   0.9965  -8.4865   6.1070   6.3619 -10.1281  10.8424
#>    0.6714   9.7294  -3.5920  -3.0837  15.8807  12.7176   4.2104  15.0739
#>   -0.8470  -1.2352   8.3293   3.4358 -16.3778 -10.4271  12.4017  -2.7037
#>    2.9145  -4.7538   2.3845   8.4128  -0.5163  14.1839   1.2079  12.1763
#>    1.3418   8.5999  -0.7566  11.3263   8.0127  -8.6011  11.7294  11.9283
#>   -2.6172  -3.2037   6.3954   1.4615  13.3616   7.8861  -4.3656  -0.7430
#>   -0.6660  -1.9773   4.1702  13.7812   7.4104  -0.4883  12.4214   8.5630
#>    9.3331  -3.4319   5.3658  -7.3925  -6.8611   1.6544   1.1920 -18.0846
#>    3.5036   2.5148   0.1629  -2.2901  -8.6481 -11.1229 -17.2549   0.5392
#>    1.9846 -10.0628  11.2063  -2.2169   1.2666   8.8498   6.4041  -1.9141
#>    1.9478  -6.5286 -10.3319   8.8431  -1.4201  -5.6046   0.0383   1.9920
#>   -6.6104  -5.0934  -2.7188   1.3301   0.5251  -9.5026  -6.4144   5.1716
#>   -1.2412  -4.4611   8.4040   6.1381 -10.5258   8.4642   2.5377   5.0823
#>   -3.7876  -2.4796   0.9942  -9.1352   5.6600   6.3992  -9.1112  -2.8160
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]