Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.5119   0.7455   0.9195  -2.7584   3.1239 -12.5394   3.2725  -7.1613
#>  -12.2179   5.7662  16.8197 -12.8277  -9.5046   4.1953  11.6349 -17.1267
#>   -1.6191   3.6810  -0.4938 -11.9881   9.3188   3.8308   1.3270   6.5519
#>   -0.8810   3.3025   2.9290   3.6616  -3.9976 -17.2719   0.9719  -2.2179
#>    7.8645 -10.9418   0.8054   0.6820  -2.3085   1.0849 -10.6534  16.5724
#>   11.2413   8.5246   9.6805   7.5028  -2.3887  -0.6417 -17.7665 -19.4197
#>   -3.8915   7.0474   3.0894   3.6957   8.1247  -7.5860   4.4420  -9.5588
#>   -8.4175  -9.3447  -1.6298 -10.5385  12.4717   3.2611   6.0567 -11.4558
#>  -14.8180  10.2778 -15.4890   0.1773   5.7691   6.8591  15.9389   8.8488
#>    8.4669  -1.7080  -0.2545   4.9095  -2.2509  -0.4705  12.4053   0.1334
#>    6.4438   2.6035   6.8866  -3.8252   5.3491   3.7314   2.5245  16.9476
#>   -2.9244 -10.0556   3.6614  -1.2904  -9.7769   2.4396   6.5331  14.7370
#>    1.9857  -3.0469 -14.9109  -3.2696 -25.4242  -0.1474   2.5668  11.2766
#>   -9.6344  -5.6774   9.5307  -2.7183  -3.3773 -24.1179  21.4858   3.4239
#>   -2.5275   0.7737  -9.0578   0.6790   9.3175  -8.4320   2.1873   1.8823
#>   -8.7048   8.2153   8.2507   5.1295   8.8091  -9.8236  11.7083 -27.1656
#>  -11.7772 -10.3874  16.5721   5.7907  -3.8616  12.8760  11.7957   8.7489
#>    6.0475   3.2605   3.9956   0.5570  -7.7775   4.9836  -8.8186   7.1853
#>    2.3587  -0.8871  -0.3382  -2.5848  -0.7278 -14.2649   3.8740  -0.7812
#>    1.8626   2.9898  -1.7714   3.3520  -0.0275   3.2399   3.4644  -6.7049
#>  -12.9546  -7.9641  12.5774 -11.2796  24.8220 -15.3271  -4.3230   1.9671
#>   -9.3138  -7.0682  -3.3725   3.2579   2.1194   7.1930  -6.8560  10.0294
#>   -0.4974  -1.4687   8.7586  -4.3031  12.2147  10.8620   3.9203 -23.7923
#>   -8.3930 -16.6531  -1.2181  16.3572 -10.3677 -17.0788   0.6950 -24.4772
#>    1.2508 -14.4939   1.3350 -16.6285  21.5171  -1.9062 -24.7435  -0.0366
#>   -4.1563  18.6449  -9.3429   3.7551   2.5912  16.8771  -7.6505   9.6402
#>   -3.2509   2.7261   1.2682   2.5222  16.9120  22.9696   7.2236  -7.1863
#>    0.7664  -2.1310  -7.3303   4.5261   3.7818  -0.9214  21.4272  -5.4534
#>    4.8670   7.3399  -1.6823  -8.7030  -5.1293   2.2450  -2.1746 -11.7518
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]