Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.3866   1.2688   2.8437  -4.4405   1.0221  -4.1005  11.5012 -19.7737
#>   -4.0517  -1.8168   3.3996   9.3336   6.5110   0.6247   4.5197   1.3017
#>    5.4335  -0.6739  12.1150   1.5434   2.9110  -3.1616  -4.1824 -11.2213
#>    3.1277  -3.6899  -4.8532  -5.0780  -0.8707  -4.2826   4.6277  -7.4727
#>   -1.0194   3.4057  -8.4607  -9.9210  -3.4217  11.4327  -7.3164   5.9119
#>   -3.6343   1.4313  -8.8614   6.2509   2.3029   1.1270  -7.9690  -5.1611
#>   -2.3263  -1.5996   3.2754   6.4089  -2.1943   3.4369   2.6662   6.2007
#>    1.9288  -1.5417  10.4305  -1.2926  -0.5738  14.6749   9.9691  -5.7059
#>   -2.8489  -3.6154   4.7237   5.4326   4.0445 -11.2147   9.5059  -4.1867
#>   -3.8121   7.0179   6.1483   4.5539  -5.0789   0.6186 -20.2172  -7.9728
#>   -0.6978  -0.6093   1.7967   8.6384  -5.7239   4.6197  -9.2259  -3.8841
#>   -5.1993  -3.1204  12.2288   9.3582  15.3618   4.3264   0.6599  -0.5883
#>   -5.7590   5.2556  11.1055  13.5218  13.7270   3.2563   0.2897  -1.9993
#>    0.4485   0.8072  -6.3411  10.4316  -7.0913   2.5939   7.7304  -1.6623
#>   -1.4840   2.0490   3.1971   1.6292   1.9189   0.3503   7.1153  -7.2150
#>   -4.8378  -3.3675 -10.4287  -2.7002  -0.6019  17.7074  17.5199   0.0534
#>    2.0078   2.7304  -4.3402   4.6320  18.0439   5.0290  -0.3847 -11.1977
#>    2.2747   7.9875   8.9310   0.3619 -12.9441  13.4214 -10.6280  11.9670
#>    1.5331  -4.1792 -14.2122  -3.9760   5.9048 -10.0162   8.1522  -0.6013
#>   -2.5233   8.3337  -3.2039   6.5475  -6.6939  14.9933  -2.8962   9.9427
#>   -4.6107  -1.2356  -4.0658  -2.7071   5.5410  -8.4771   8.4540   2.0024
#>   -3.9098   0.3080  11.6518  -5.4028  12.6345  -8.5591  -4.8900   6.1889
#>   -0.3843   2.9100  -1.0902  -5.8782   1.6696  -4.9083   1.1965  14.3514
#>   -3.8995  -3.6090  -1.3649   4.3000   9.0932  12.4956  -4.3375  12.6740
#>    3.3480   0.5410  -7.7530  -6.7069   0.6505 -10.4505   8.1586   6.2072
#>   -1.9056   1.7619 -12.7043  17.0078  -7.3784   3.4692  -9.9376   3.9898
#>   -2.6427  -1.7192   9.8266  -5.7919   1.4813 -21.0357  22.8381  -1.4001
#>    0.0046   8.8437  -7.9392 -13.3916  -6.1309 -13.7265   1.9363  -6.5925
#>   -5.6157   6.1119  10.9169  -3.7167  -6.4288 -12.8373   8.0737  -5.9788
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]