Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.7050   5.5253   6.4420 -10.3322   8.8184   8.7760  -3.8392   3.5731
#>    6.9510   3.8679  -7.4144 -11.4203  -8.5823   7.8983 -12.9172  -1.8452
#>   -2.4685   3.7456  -1.0900  -2.2860   7.4590   2.5840   1.0120  13.2445
#>   -7.8924   1.5603   9.9072   8.5869  -5.4033   1.0779   3.6815   4.6639
#>    1.7750  -2.8109   2.6242  -0.8754  -5.4098   1.8248 -12.4309  -3.1932
#>   11.6846   0.1857  -8.9932  -7.5093   6.1973   7.0161   2.6129   1.8319
#>   -1.2647   2.6036  -4.1753   4.3465   3.9162   4.4206  -2.8990  -7.3588
#>   -0.7876  -5.6109   1.1174   8.3624  -5.6245  -5.2037   2.4479  16.8769
#>    3.6673  -5.0606  -8.3507  -7.2014  -6.4115   3.6778   9.0660   1.6938
#>   -1.7507  -8.1032   2.5156   0.8912  -7.1661  -1.8206  -7.0282   3.1515
#>    2.3510  10.4812  -0.6240   0.3996   5.3944  -6.8760   3.8437   2.0440
#>   -5.0577  -1.8202  -1.3119   8.5769   7.7899   0.7959  -2.5136  -8.1623
#>   -0.5888  -3.7809   5.0724   3.7380  -1.1854   0.9871   8.6404  -5.0669
#>    0.2994  -0.7468   2.9127  -4.0931   5.3563   0.5474 -17.8929  -0.5042
#>   -5.7441   7.7318   9.2955  -4.5509  -1.6290  -3.9527  -3.1748   1.4204
#>   -4.9397   8.4134  -0.6814  -4.8818   5.1993   8.7330   2.5912   1.3081
#>   -0.2147   2.4817   5.3563  -4.8975   7.2386  -3.7926   1.5411  -2.4183
#>    1.5672  -3.8165 -12.8196   4.9598   5.1569   8.8311  -3.3130  -9.9431
#>   -2.4574  -4.6099  -2.6097   8.3215 -10.8287  -8.6996  -4.1054  -3.4209
#>    0.0876  -7.6342   9.9528  15.9554  -9.4757   7.4745   7.4369  10.2211
#>   -2.1012 -13.0096  -9.9969   3.5695   6.9946  -3.1968  -8.4315  -1.9118
#>    2.4987  -0.5681  -4.6548  -4.4445   0.3194   3.2404   1.1623 -12.4677
#>   -1.2406  -6.9182  -3.8939  18.3968   2.9119 -11.6019   6.4638  -2.4888
#>    3.9243   1.7727   3.2870   1.6940  -3.8129 -13.5767  -3.7639   2.3852
#>  -10.1023   0.1540   4.5039  -5.3617   1.1896  -2.8297  -0.0411   4.7370
#>    7.0295   4.1925  -9.5464  12.5397  -3.8125   4.7275   3.5714  -5.7386
#>    7.2237   6.4135  -1.1159  -3.5016   3.4269  13.8637  -5.2271 -10.2839
#>   -1.6404  -6.1120  15.9230   5.2491  12.0508  -5.2522  -2.8307  -2.2380
#>    3.5463  -1.4269   0.8965  -6.1280  -4.5990   4.2573   4.0617  17.3347
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]