Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -9.0841 -12.2781  14.9447  -8.7966   3.4183  -6.5102   0.2356   0.8004
#>    1.6508  -2.4323   8.8984  -5.5236   9.0438   7.9622   9.1304  -2.1583
#>   -1.6674 -13.6943   2.8203   5.4945  -5.7256  -7.7610  20.0388   4.2691
#>    6.6599  -3.1332  -1.7308 -10.9023   5.9605 -17.0757   7.2569   8.7708
#>    4.0788   2.1244   5.2517  -4.9727   5.3405   0.3329  16.5991   8.8711
#>   -8.3913  16.0850  11.0290  10.8549  -0.6191   4.5706  -2.7801 -19.2022
#>    2.1014  -2.7310   2.3903  -3.0074   2.2040  -0.4331   4.8830   3.8976
#>   -1.0673  -0.2272   4.5455   0.6964   7.5585  -5.1316  -7.1794 -19.5935
#>   -1.6255  -2.5206  -1.8966 -12.3350  -6.9871   4.0577   6.7168   3.9232
#>    2.2444   3.0828   5.5755   1.7330   1.5761  -4.1851   9.9794  -0.8612
#>   -0.1028   0.7083   3.7323  -5.6333  -6.1871   8.6304  -2.3643  10.7087
#>   -5.9431   3.8464  -8.3758   2.4482  -5.7363   0.9986  -6.0517  17.7035
#>   -8.8032  -4.4070   0.0908  15.4200   0.1858   8.1170   5.9593  -6.0738
#>    2.1718  -2.0452   7.8162  -4.3622 -12.1343   0.3852  -5.3287  -0.5992
#>    1.3729  10.4519   5.4567 -20.3376  -0.6751  -3.1473 -23.4498  -7.3894
#>    3.3535   6.2391  -0.8513 -14.6386   9.4787  15.8923   3.7523  10.5790
#>   12.0973  -6.8440  -6.2186  -3.1421   6.9182  -6.0678  -8.5062   3.2622
#>    5.7212  -3.9400   8.4999 -11.6282   2.2355 -21.6639   0.0196  10.4879
#>    9.0044   2.4175  -1.2692 -12.5713  -9.1179  -2.0845  -3.3951  -3.1546
#>   14.3545 -11.2021   0.2925   6.9321   3.4116  -3.5444   6.0039   1.2575
#>   11.3728  -0.6044 -10.1516  14.2797   0.9271  -1.6928   2.1214  -5.0738
#>   -0.6975  10.1900  -0.3302   6.0117  13.3342  16.1527   1.7033 -10.1698
#>    2.4385  -0.2405  -4.1776  -2.9993  11.4642  -3.8872  -2.3654  11.8486
#>    4.4158  -4.8250  -5.9348  -4.0684  11.7630  -0.8902  -9.2298  -3.8429
#>   -4.7263   2.0880   7.9309   9.8904  -6.3356  14.9696 -20.2938 -13.1081
#>    4.6491  -2.5497 -10.4439   4.8228  11.3711 -11.1134  -2.6513  -5.3213
#>   -2.2101   2.4101  -2.7072  19.7530   8.2281  -2.2477  10.4118   8.6362
#>    2.7541   6.8084   1.8346   1.5774   2.6468   1.2345 -13.1184  -5.2807
#>    0.9217  -3.5848   8.2527  -6.8506  -7.0980  -1.7071  -0.1172  -3.3970
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]