Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.1558  -0.0221   0.4488   1.5226  -1.4436   2.1834   1.6219  -4.6328
#>    5.0554   8.3189 -13.9535  -5.9473  -3.3695   1.4616   9.3747  14.3166
#>    0.1432  -1.1098  -3.3916   7.7032  -2.8518   2.6429  24.7834   9.0417
#>    1.5646   0.6535  -1.6723 -10.7368   9.6706 -10.4836  -6.1284   2.5089
#>    7.9648   7.7468  -4.9403 -11.7508   4.0173   5.7061   6.4023  -2.0814
#>    0.8192   2.3250  -6.2881  -1.6584 -12.7731 -17.4993  -9.1124   3.2166
#>    2.3720  -4.0594   0.7087   2.9156   2.4721   6.0874  -1.9855  12.9209
#>   -3.2055  -1.1335   9.8091   6.0041  10.3410   5.4807   4.5985  10.7206
#>   -1.0381   1.9136  -2.2830  12.1873 -13.9264  -1.7116  -5.9406 -11.9636
#>   -0.4606  -1.4432   0.8515  -7.1805  -4.2104  -1.9473   3.9360 -21.2906
#>   -1.0873  -4.4121  -0.1391  -3.0880   5.8651 -19.1126 -10.2547 -14.4272
#>   -2.1640   7.5272  -2.5500   1.0673  -4.8790   6.1969  -9.7611  19.9481
#>   -1.9555  -7.9014   7.8417   9.0437   5.5558  -4.5404 -14.1764 -16.5829
#>    9.2861  -5.2191  -7.3230 -14.5258  -3.1968  -7.8510   8.0626  13.9893
#>    6.1021  -6.3451   3.8570  -1.5956   5.2669 -12.6247   8.1593 -21.8336
#>   -3.3622   3.3446   6.0914  12.2190  -0.6543   5.4972   2.5584  -3.5543
#>    4.1210  -1.8048   5.9687  -8.0158   1.2846 -12.3115  -5.9355   6.3404
#>   -1.0861   2.4353  -3.6344  -5.4193   1.5530  -7.8990   8.3212   8.4361
#>   -0.0437  -2.2984 -10.9248 -10.9382   4.0639   3.6512  10.3909  -3.6629
#>   -7.5662  -9.9131  -3.9683   0.7312   4.2574  -1.7508  -8.4690  -4.8365
#>   -4.7571  -1.9577  -3.8623   0.1185  10.1148 -14.9303 -15.6584  10.2452
#>   -0.1259  -1.3342   4.0466  17.4495 -10.2009  -1.7148   2.8152   8.5092
#>   -1.0115   5.8274   5.3891   3.8811  13.7145   1.7384   1.4977 -17.1632
#>   -1.3821  -2.4043   7.3472  -1.1787 -15.4592   6.8963  -0.7502 -18.1407
#>   -0.6462   5.5767  -4.3102  -5.6653   1.8414   2.2718  -3.5502  15.1789
#>    4.5993   6.8616   2.7664  -3.4739 -17.4600 -10.1429 -18.6338  -2.7327
#>   -7.9589   1.8941   2.4878  -0.1396   9.4732  -5.3607 -16.3696  -1.4702
#>   -1.0211   0.7450   3.9914   2.4361   0.3933   4.3494   0.6884   4.9420
#>   -2.0266  -2.0878   7.6878  -6.6572  -2.6840 -13.7956 -12.9463   6.7778
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]