Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.1353  -2.5971  -0.0935  -1.4145  -4.6087   1.7325   0.9498  -7.0461
#>    0.4050  -4.0127 -11.1108   9.8861  -9.6073   5.0468   8.9159   3.1428
#>    0.1182  -1.7248  -1.9387  14.3486  -9.2990   8.5348  -6.4173   9.0199
#>    1.0858  -4.5051  -0.7951  -9.3093   5.8883   6.0174  -9.7963   4.0807
#>    5.0338   4.2420  -1.1746  -3.0585   4.3703   4.5833   7.0847   4.2242
#>    1.1624  -4.6851   2.9668  -1.3512   0.8447   0.2881 -11.9045  -8.5977
#>    2.5735  -0.6615  -1.2756   5.7620  -7.8336   1.5098   0.0658  -9.9994
#>   -0.9093   2.3373  -1.9338   4.0168   8.6603  -9.1380  10.0785   7.1205
#>   -1.2672   2.5253  -2.7692  -4.1302 -11.4272   5.4788   6.3121  12.1103
#>    6.0802   9.4527   3.5726  -9.2597   7.7919  -0.1448   0.0977  -7.0710
#>    3.9643 -10.9274   9.5669 -10.7886  10.1763   7.0614  -1.3338  -4.0032
#>    1.6289  -0.6552  -3.3093 -18.0983 -22.3189  -6.7895 -22.4300  12.8340
#>    4.0510   0.9513  -5.1168  -1.4318   0.7165 -14.0180 -12.1255  18.9426
#>    0.0264   5.0460  -2.0239 -10.1038 -19.1143  -9.9204  -8.4888  -7.1236
#>   -0.9439  -3.9506   8.0023   4.4858   4.1432  -1.3229   4.6247   0.9467
#>    0.2956  -0.6230  -8.0244  -7.6212  -2.9728   5.5669   6.6473  -2.1877
#>   -4.3614  -3.7734  -0.3783  -2.9192  18.7633  -1.2798  -1.5500  -0.9722
#>   -5.6475   8.4090  -2.6160 -11.8978   0.1499   7.2446   3.9341  -1.2463
#>    3.0728  10.6079   3.7898   1.0177   5.2827  -0.8679   3.4547   5.8709
#>    2.4273   2.0509 -17.1700   9.1115 -18.0229   9.3479   1.8241  -2.5298
#>   -5.8778  -4.5635   1.1686   2.3346  -8.5575   2.4384  -0.0220  -2.2903
#>    0.1431  -4.7390  13.3413  -1.6572  -0.9111  -1.0786  -4.9531 -15.0147
#>   -7.5496  -2.4998   6.4993   2.5850  -1.8604  -2.3187  -7.5389   5.5005
#>    3.7870 -11.0804  10.0669  -4.0193 -10.5778   7.3933   1.0191   0.8948
#>    0.9252   7.5999  -3.5341  13.8511  -0.9643   2.9260   3.6168  -1.6100
#>   -3.1650   9.0298  -0.2275   5.7272   2.2128   6.0943   3.7215  -9.9676
#>   -2.3863   6.4056  -6.3203   5.7825  -5.7782  -9.2225   7.0808   6.8668
#>    2.7919   7.8078  -5.6275   4.5158 -12.2972  18.6311  -9.6907  -0.4494
#>   -5.9563   2.7402   0.6764   0.7702   6.4487  -2.3495  11.7819   4.9748
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]