Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.0573   4.2197   4.6499 -15.2861   9.3790  -3.0689  -1.2035  10.9293
#>    3.9972  -2.1034  -4.4743  10.6302   2.5307   3.7956   1.2784 -13.2556
#>    3.2990 -13.7689  -0.8004 -16.2704  19.6313 -10.1412   0.2267  15.7822
#>   -1.2677  -0.4232  10.6098   5.9832  -7.1119  -3.4594   4.1388   0.1196
#>    2.7822  -1.5979   4.6833 -14.4667  10.5714 -16.1614  -7.4061  18.3708
#>    0.3528  -3.7546  -2.8913  -5.8873   4.0283  -5.9494  -0.1330   5.1909
#>    5.3788  -6.8221   5.7717 -11.2996  18.0000   0.4134 -13.3762  -1.1963
#>   -0.3845 -10.7886   9.9237   7.5735   2.1423 -14.4470 -21.8758   4.3481
#>    3.8320  -5.5311   3.0872  -1.4091  -1.7240   7.8842 -10.1582  -4.5623
#>    1.6686   3.0945  -3.8649  -1.9452  -2.9407  -9.9856  20.3220  -3.0006
#>   -0.8279  -2.9561   3.2578  -3.6590  16.7125   2.0995   1.3049   4.8144
#>    2.7572   7.9335  -0.6211  -3.3610   5.9329  16.2848  -4.5972  17.5318
#>    0.7943  -0.2937  -4.5202  -9.7346   5.0631  14.5867   4.6325   4.9348
#>    0.1504  -1.0030 -14.4147  -1.2423  -9.5477   7.3190  -4.4228  20.0052
#>   -1.3820  -7.5952   0.4119   3.3336  20.1217  -3.5948   8.0489  10.8160
#>   -0.4199  -2.1809  -3.8093  -1.2865   5.1646  -1.5133  15.5693   0.1545
#>    0.8772  -4.3040   2.5808   0.6129 -21.0089   1.0073 -11.8728  -6.3413
#>   -0.4816   1.4241  15.9402  -5.7808  -2.9679  14.5103   7.8121 -12.7104
#>    0.7161  -3.1126  -8.1791   8.0053  -0.0671  11.7388   0.9553   8.2901
#>   13.4437  -5.3178  14.7850  21.3245  -6.6561  25.1181  -8.3551   1.6861
#>    0.8829  -6.0993 -13.6619  20.5098 -12.1224 -13.0113   0.3260  11.5021
#>    0.8621  -3.1730  12.8408  -6.0280  -2.9823   2.2610 -14.4290  -2.0979
#>   -0.1606  16.3261   2.5959  -5.4684 -10.5681   6.4644  -3.1605 -18.8069
#>   -3.4506   3.5140  -2.6810  -3.2469   3.9745  13.1055 -11.4575   6.3343
#>   -5.6997   1.5887 -20.1420  10.9489  12.0659   1.4387   4.4189 -15.9829
#>    1.9570   0.7672 -11.4439  15.1088  -3.0707  23.1829  -4.2660   5.4326
#>   -3.1846  -6.1663   3.7927 -18.4117   1.7420 -16.8749  -1.9940   9.3169
#>    5.2762 -18.2832  23.3502  -2.0166   6.5767  -8.8203   9.6708   1.1231
#>   -2.6850   4.7898 -10.3529   6.8311  13.1598   3.9428   6.9617  -2.8546
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]