Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -6.8063   2.2600  11.1844   1.6174  -6.5552 -22.9088  -2.1310   5.6289
#>   -1.9310  -2.2586   0.4749  -1.1708   9.6432  -5.9833  -3.6570   8.5729
#>   -2.6016   7.5028  11.3450 -14.2609  -4.4834  -4.4104   9.1064  -8.4013
#>   -4.8491  -3.9126   5.8663  -0.8049   3.9233   3.2368   3.2296  -5.4614
#>   -2.2962  -4.6887 -15.7214  19.6600   9.4913 -12.4114   2.8877 -13.3137
#>    0.3702  -6.8425   3.5405  -3.2479   3.2414  -2.8500  -2.8128  -4.7773
#>   -5.3514   1.9286  -4.6502   6.3450  -0.0520  -6.5854   7.6441  -0.8781
#>   -1.1791   3.6521   0.0393  11.8783   7.4889  -0.4242   0.5687  -6.4037
#>   -0.5261  -2.1269   4.2076   1.6008   7.5667  10.7329   2.5983  23.0423
#>    1.6764   1.8108  -0.4834  -8.4429   2.6526   1.8694  -1.9265  -8.1984
#>    3.7547  -0.0224  -6.2305  -1.8044  10.1340   6.4819  -1.4973  -7.0727
#>    6.6253   7.3566 -16.8927  -0.3007  -5.3435  11.7457   2.2335 -15.7880
#>    1.2732  -3.7002  12.7204   4.7625   4.8267   2.4135  11.4489   1.0281
#>    1.6743   0.5900  11.7818   6.6038   0.3594   5.8780  -7.2219  11.3949
#>    1.7496  -5.7326 -12.0977   5.2370  19.0627  -1.8326   2.1200  -4.8898
#>   -2.9149  -2.2502   8.8985   2.6612  -4.7317  11.7948  10.2038   1.3824
#>   -5.1822  -2.9069  13.6341   9.3046 -12.0108  -0.9739   8.9688   7.2614
#>    5.3822  -2.4326   2.3028   4.0593   8.8223   6.9071  15.7676   0.1736
#>   -2.3625  -1.0499  -9.1793 -17.6451  -6.7444  -6.0926  14.6475  14.2448
#>   -1.2228   1.8301 -16.9652  -9.1019   4.8187 -12.7384 -16.7558  -4.2111
#>    6.9005  -3.8407  -0.7601   9.6976  -6.6935   0.8202   1.7050   0.7569
#>   -2.1741   2.9772  -4.4049   8.2494   1.1030  -7.3431  -6.8110  -5.2426
#>   -2.8983   2.9344   7.7024   1.0897  -5.1812  -6.4622   1.3201  -0.3040
#>   -4.6536  -5.1943   0.5211 -17.2616 -10.5155   4.7127   0.1779  -4.2046
#>    6.3586  -4.6635  -3.7396  -5.8778   9.5452 -14.8455   8.9324   2.6500
#>    5.5172   4.4068  -2.4458  -8.4581  14.9992  13.9129  21.6846 -16.5633
#>   -1.2351  -0.6564   8.5328  -0.2910  -4.5482   1.3941  -7.6644 -17.9504
#>   -5.5271   2.7164   7.3947  -7.5018  -5.7027   1.7437  18.0244   5.8862
#>    1.8384  -5.3274   1.4605  -0.0236   8.8169 -11.1560  -3.9956   5.9900
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]