Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.2084  -5.8141  -0.8801   1.2605  -8.6953  -0.8669   3.6650   6.5223
#>    3.3127 -12.1492  -8.6004 -10.2101  -1.5990  14.3296  14.1451  12.6026
#>    3.3554  -0.9924   4.7163   6.1382  12.0637   5.2409  -6.3697  -4.3751
#>   -1.4085  -8.1366   4.6712   7.5977  -9.2217  -3.9457 -15.3311  -7.1434
#>   -0.6285   0.7180  -3.6482  -4.5328   2.9891   4.9871  -6.8022  15.2074
#>    1.1333  -4.1839  -0.4656   6.8492   6.7442   9.3658 -14.3487 -21.6214
#>    2.6998   0.6238  -6.4687  -9.9976 -24.3201  -4.2562   2.2143  -1.6482
#>    4.4625   3.9326  -0.6564  -3.2671   2.4350  13.9899  18.5328   7.1094
#>   -2.8222  -2.2168 -10.2529   8.3583  12.9500   1.9326   7.4026  12.3720
#>   -2.8686  -0.1358 -16.3651   0.6156  10.7123  14.9732  -3.1658  30.0341
#>   -8.4777  -1.2903   0.2652  12.2376  16.6139   1.1661  11.3519  -4.0367
#>    0.6961 -12.0772 -11.9015  -3.9783   4.3665   8.5450  13.3732   4.0459
#>   -2.7248   5.2083   2.7694   7.6754 -11.3604  -2.0812  -1.3399  -3.1469
#>    2.8966  -1.9115   0.1254  -1.7730  -4.6926  -8.0471 -17.8863   5.4684
#>    8.3475   9.5719  10.8115   8.9925 -10.1631  -7.5961  -3.9896   2.0238
#>    3.8589   6.6410   8.3678  -3.4333   3.2108 -19.6612  -5.5888 -11.2914
#>    2.6105   2.7534   5.9227   1.0781   4.7186  13.9410   2.4931 -11.0315
#>    2.8726  -4.3017   9.2225  -1.2166  -2.3600   5.1744 -14.3057  -8.8461
#>    1.3744  -0.3748  -0.5653   6.0596  10.1486   9.6968  -0.6417   7.8065
#>   -3.6333   2.3057   4.7845   4.0186  25.2289   5.0552  -4.2294  -9.4497
#>    3.3483 -19.2685 -10.4513  -4.6746  -8.7733   0.7014  -2.6900  -5.9106
#>   -2.4565  -3.7501  -5.5801 -12.5271   1.0953   4.4772  -2.3515  -2.6530
#>   -5.1629   4.2736  19.8736   9.3838  14.9250 -15.8973   6.2226 -30.0260
#>   -2.6015  -1.9405   4.1343  -8.4327   8.5529  -1.3428   6.8600   7.1543
#>    7.3030   7.5481   8.1050  12.6037  -4.0135  -6.1742  -2.3783  -9.9678
#>   -3.8909   4.9858   0.1312   4.8430  -6.8866  -5.9146  -8.2320  -1.7664
#>    0.9021   0.6016   2.5444 -15.2325 -15.8297 -10.0461 -24.4232  -2.7698
#>    0.9765  -3.2103  -2.7951  -2.1233 -10.7681  14.9798  -2.6062   3.3115
#>    3.4443   1.3257   5.7150  -7.6240  -9.2673  -0.4160   2.2430   7.4261
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]