Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -8.7647  -0.4682   7.7963  -4.5369   9.4942  -4.1617  -4.1103  19.3087
#>    4.0360  -9.8447   0.9105  13.8762 -17.0392  -2.3065  13.8410 -10.6713
#>    3.7199  -5.4277  -7.3849  -2.0523 -14.1921   0.8237   2.7750 -16.7750
#>   -4.5997   6.8190  11.9580  -7.3908   0.7623  -6.0993  -7.8714  -0.3275
#>    6.3350   5.6915  -0.2486   4.5125   6.4556  12.5466   5.1221  -8.3328
#>   -5.3670   3.7056   4.6091  -5.5202  16.5126 -14.6716   7.2231   6.2862
#>    0.0741   2.0996   4.3167   4.2941  -0.6075  -1.2789   2.7319  -3.0481
#>    1.4276  -3.3137   2.0153  17.5720   5.8865  -7.6347   0.3849   9.1147
#>   -4.3790  -1.2413   2.5337  -4.7801   2.7141  -3.4421   7.6487   2.9184
#>   -4.7217   0.6327  13.1321  -4.9202 -15.7671   8.1676   1.3092  -7.3948
#>   -6.7910  -9.0685   1.4939   9.2089   9.1012 -12.8364  -2.4851   8.3031
#>    2.3443   3.5523 -12.5885 -13.3636  -7.0710  -2.8443  -5.6563 -17.1588
#>   -2.8623  -1.8491  -4.7443 -10.1820   4.6370   6.6014   2.3184   0.2719
#>    1.5113   2.9705   0.9446   1.3710  -4.9231   1.6559  21.6072   9.9877
#>   -0.7059   2.8873 -10.8913 -11.6808  -1.1437  -3.4593  -6.2159  -4.9284
#>   -2.9788  -2.5528   1.3435   8.7394 -15.1196 -16.0828   4.2766  -4.3610
#>   -1.2020  -7.8141  -0.9332   5.9856 -18.5626 -15.4667  14.2923  17.6954
#>    6.0894   4.3110   0.2822   0.1223   1.3098   0.9189  -3.4149  -2.1252
#>   -1.4548  -1.2519  -5.2414 -13.8979   1.0132   3.9838  -5.8436   1.6588
#>    2.0508  -1.8697  -7.5357   1.9029   7.8309  -5.8733  -5.8976  -0.4331
#>    6.5930   7.0586   7.5887   4.3415  11.8226  17.3290 -18.6833 -12.3254
#>   -3.0258   7.2572  -0.1697 -15.1733  -0.9425   0.1767  -6.7976  -7.8783
#>   -3.5671  -6.8015  -0.3536  -2.0729   1.7173  -6.4490  -9.9026  14.1454
#>    2.0407  -4.4729  -5.9910  -6.2875  -1.3764   5.0817 -20.3692   5.0212
#>   -2.0986  -3.7933   3.6157  -0.7303   5.8362  -1.0007  -4.7152  -0.4602
#>    1.9821  -3.1704  -4.1633   2.1748   4.3099   8.9313  10.0927  10.6518
#>   -0.0687  -3.4093   9.0262   4.9696  -0.1656   1.7001   9.2772  -6.3854
#>    4.2472   1.8456   3.7126  -7.6332   0.8242  -7.2192  -6.1020  -2.6523
#>    0.6477   5.2059   5.2467  -2.5625   5.8792  -8.6343  -8.8740  -3.7354
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]