Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.3729   1.3786  -9.2917   3.4004 -11.7277   4.3711   8.2306  -2.9114
#>   -3.6327  13.9869  -1.5534  -6.6108  -2.8981  -3.6020  11.9620 -16.2006
#>    0.0822   9.4913   9.1161  -5.2808   2.9217  -0.8371  -9.1493   4.4485
#>    0.7385  -3.3810  -0.0333  -0.7891  10.5992   3.6477   8.7140  11.2261
#>   -0.4761  -1.1888   0.1147  -1.0332   1.0878   3.2406 -10.0694  -3.4029
#>   -3.1179  -0.6978  -3.2251   3.5631  -5.4007   0.2074  -2.2534   7.2827
#>   -1.1774  -1.6160 -11.3457   7.6584  15.0447   5.1768  -8.1084   2.0209
#>   -0.9703   1.7024  -1.6035   6.9381  -2.5021  -7.1567  17.3173  -7.3277
#>   -3.5720  -3.0593   2.1026  -5.2734  -4.8600 -10.7559  12.0979   2.7051
#>   -4.1633   1.4250   5.8134  -6.8451   7.3816  -4.2849   1.0620  14.9210
#>   -6.0330  -3.4921  -5.4017  -7.2411  -9.3450 -16.9198  -4.1521  10.6309
#>   -0.1570  -2.0075  -4.4718  -3.7954   1.6623 -11.5174  -1.4869  -1.0195
#>   -9.7662  -9.0853   2.9720   3.9667  12.7465  -7.3985  -5.1837  -9.4741
#>   -7.0187   9.8752   2.6333  20.1178 -12.4809 -13.5843   8.4565 -13.5424
#>    0.0828  -1.0795  -0.0465  -1.1791   2.8202  21.4192  -9.9863  -0.5543
#>    4.4308  15.3921  -5.5802  -5.2437 -11.7881   4.2197  15.3974  -2.5088
#>   -7.6187   1.3727   5.7866   2.4860  12.9224  -4.5247 -10.4235  -9.3233
#>    3.6361   1.2460  -1.0704  11.9619 -17.0932  -2.1554   1.4222  -3.3986
#>   -6.2627   4.1258   9.0407  -3.0431  -2.0206 -19.2288 -11.2199  -0.6557
#>   -4.3547  -2.6458  -4.4549  13.2301  -2.0056   4.2969  -5.1472 -12.4104
#>   -1.9505   1.2553   6.5993  -1.8541  -3.0898  11.0779  14.6180   8.5317
#>   -1.4253  -3.8969  -8.4487   6.0993  -3.9375  18.7791   2.7184   1.2481
#>   -1.2800  -2.8705   3.3551   5.3958  14.0499   0.4978  -0.1509 -10.5845
#>   -3.2118  -4.1342  11.7434   0.3969  -8.3439  -9.4905  -6.4631  10.5380
#>   -4.7966  -3.9950  -6.6500  19.5838   4.9869 -14.7302   0.6593  10.4676
#>    2.2725  -0.3695  -2.6248  -1.2067  -9.4904   4.0800  -6.4005   2.1096
#>    0.9604  -1.3836  -0.9625  -5.7691   7.8579  -1.0361  -7.4859  -9.6930
#>   -5.6554   0.0568 -15.0160   4.0324   1.5215   6.3651   3.5131 -14.2811
#>    7.9878  -3.2400  -6.7029   7.5199  -4.2193   4.7146  -2.9976  -2.1031
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]