Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.1887   0.7418   7.8526   1.7516   1.8723  -7.0970   0.8289   7.7919
#>    5.0371   7.1776   4.5895  -5.5314  -0.5328  -0.3061  -4.6702  10.4968
#>    1.2801   5.4034  -8.9779  11.9492  -3.9221   0.4663  11.3811   0.3541
#>    0.6197  -1.6282  14.8515  -6.8339   7.7621   1.8312  10.5831  -1.1812
#>   -1.4367  -0.8093 -10.2703  -1.2518  -5.9560  -9.0067   4.3638   9.0887
#>    1.2133   2.7113  -0.9504 -13.9826   6.9656   7.8086  -5.0170  -7.3714
#>   -5.5831  -3.2706  -9.5675   9.8038   3.3016  -1.6436   1.4007 -10.7597
#>    5.5400   4.2999  -1.7382   1.8209   6.7655  -0.8509  22.9715   5.7545
#>    7.1577  10.3714  -3.0483 -10.4886  -0.9990  -1.7891  -1.6825 -10.2044
#>   -1.4200  -1.1979  -2.1541   6.9658   0.9810   2.2290  -1.0892  10.8145
#>    0.0631  -6.0745  11.6390  -5.9297   6.5977   7.6619  -7.8140  -1.4879
#>   -2.5386   2.1901   2.0057   4.7995  -8.6698  -2.0942  -9.0016  -7.4355
#>    1.3076   0.3756  -5.8489   1.6724   2.3502   1.5638  13.1823  -6.8015
#>    0.4062   8.2306   2.5287   9.6685  13.8220  -9.9091   8.6520   1.0282
#>    3.0365  -2.8734  -3.1408   7.7913 -14.5415  10.1435 -19.4883   4.0986
#>    1.7823  -1.2794  -5.0009  -6.6691   0.8091   4.9558   7.7604  12.5935
#>   -5.3990  -4.3337   1.8056   2.3026   8.9945  -9.1206   7.7696  -1.3390
#>    2.3621  -6.2352  -3.2801   4.4524 -10.3302  -3.4068  -0.9963  -2.7593
#>   -1.9449  -0.7997   1.0213   2.6817   0.3732  -3.3087   3.2975 -12.0129
#>   -1.6838   3.4483  -4.2622   4.2457  -0.7063  -8.4358   7.2876   1.8726
#>   -6.2198  -3.7876   6.2870  -0.0016  -1.8208  -0.7753  -5.5594  -6.5687
#>    4.7400   1.5398  -0.9491   3.1184 -11.6237 -23.5271  -8.3505  -1.4371
#>   -4.9820  -7.9426  -5.3403   1.6124  -4.0723   5.3572 -11.0434   3.1692
#>    0.7189  -4.3446   9.7899  -4.1517   5.6621   4.3441  -3.8980  -4.5156
#>   -0.1858  -2.7913  -9.0827   9.5728   6.4583  16.1753  -7.1291   7.8337
#>    3.1210   1.4744   3.1031 -19.3892  -2.6592  -7.3025  -9.0405   9.8102
#>   -1.4275  -1.2764  -5.2934  -8.7633  16.5892   2.6531 -10.5464   1.1552
#>   -1.0674 -10.0456   7.4404   1.2960   6.6368  -8.1588 -11.0634  -1.2173
#>    0.6065   6.7794   0.2618  -2.9815  10.4455   4.0294   2.1507   7.8095
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]