Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -1.8789e+00  3.4584e+00 -3.3708e+00  2.2324e+01 -2.0484e+01  5.2699e+00
#>   4.8249e+00 -5.5114e+00 -1.7826e+00 -1.1626e+00 -7.1541e+00 -3.5333e+00
#>   6.2980e+00  4.4601e-01  8.6233e+00  5.4007e+00 -1.1125e+01  3.9077e-01
#>   1.5484e-01  2.1783e+00  4.0463e+00 -9.8054e+00  3.2139e+00 -2.4401e+00
#>  -2.6471e+00 -4.1098e+00  3.2478e+00  5.2029e+00 -1.7068e+01  1.9224e+01
#>   4.7061e+00 -4.2642e+00  5.6602e+00 -1.0119e+00 -2.5714e-01 -5.2064e+00
#>  -7.5276e+00  3.0365e+00 -1.5134e+01 -1.8930e+00  8.0873e+00 -9.3058e+00
#>   6.9240e-02  4.5393e+00 -6.1311e-01 -3.4206e+00  3.9735e+00  1.8455e+01
#>   2.1312e+00 -4.3714e+00  3.8258e+00  5.7279e-01 -1.8847e-01  5.6395e+00
#>  -4.6270e+00 -1.7695e+00 -4.4403e+00 -4.2174e+00  7.4702e+00 -5.3724e+00
#>  -3.2035e-01 -3.4575e+00 -7.3532e+00 -3.4814e+00  9.5011e+00  2.7930e+01
#>  -1.8520e+00  3.4310e-01  5.5356e+00 -7.9254e+00  1.0652e+01 -5.9294e-01
#>  -6.1844e-01  3.3815e+00 -5.5812e+00  1.4875e+01 -6.0682e+00 -8.5699e+00
#>  -1.6299e+00 -5.3660e-01 -3.5035e+00 -1.8022e+00  5.8909e+00  7.4263e+00
#>   2.6333e+00 -4.6662e-01  5.9880e+00  1.3393e+01  8.8600e+00 -7.8275e+00
#>  -1.6636e+00  2.3682e+00 -2.6957e+00 -4.4580e+00  8.1054e+00  1.1193e+01
#>   1.7303e+00  5.9525e+00 -8.3727e+00  3.4668e+00  9.9642e+00  5.8285e+00
#>  -1.5377e+00  4.4929e+00 -8.5794e+00  7.8130e+00 -1.9554e+00 -1.4887e+01
#>   2.1916e+00  7.8064e+00 -9.4226e+00  5.6598e+00 -4.1044e+00 -4.8933e+00
#>   7.3202e-01  4.6387e+00  1.5719e+01 -8.0832e+00  7.1893e+00  2.3725e+00
#>   4.4403e+00 -1.0521e+00 -1.9536e-01 -1.2962e-01  1.5733e+00  3.9821e-01
#>  -1.9336e+00 -1.7924e+00 -7.8537e+00  1.8640e-01 -6.4076e+00 -1.3766e+01
#>   6.8410e-01  2.6273e+00  1.1450e+00 -5.7058e+00 -8.1590e-01  1.8382e+01
#>  -4.8927e-01  1.1716e-01  5.1106e-01  4.5971e+00 -2.9781e+00 -1.0291e+01
#>   1.2762e+00 -1.8586e+00  4.3768e+00 -2.2809e+00  4.0790e+00  7.8380e+00
#>  -6.8246e+00  4.1073e+00  1.7249e+00  5.1775e+00 -1.6422e+01  2.4016e+01
#>  -3.2111e+00 -1.4138e+00  3.8810e+00  3.4727e+00  2.0029e+00  1.0224e+01
#>  -1.5806e+00  9.1170e+00 -6.3465e-01 -9.8058e-01  2.0090e-01  1.9834e+01
#>   9.4471e-01  1.9867e+00  6.7875e+00 -6.4958e+00  6.3101e+00 -4.8348e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]