Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -3.9579e+00  5.5987e+00 -7.5938e+00 -1.1905e+00 -9.8121e-01 -3.1968e+00
#>  -9.2529e-01  5.9325e+00 -3.5159e+00  7.1118e+00  1.0473e+01  1.2072e+00
#>   8.8107e-01 -1.0616e+01 -6.4644e+00 -1.0344e+00 -4.9115e+00 -1.2069e+01
#>   3.1065e+00 -1.0279e+01  5.5365e+00  2.0532e+01  3.9279e+00 -1.2367e+01
#>   2.5626e+00  7.3395e+00 -1.1819e+01 -1.7669e+00 -5.6693e+00  1.4497e+00
#>  -6.3763e+00  6.9013e+00 -6.2582e+00  1.5376e+01  9.6574e+00 -1.6975e+01
#>  -3.2558e+00  1.9314e+01  2.4100e+00  1.2301e+01 -4.5966e+00  7.5348e+00
#>   3.8140e+00 -7.1949e+00 -2.0208e+00 -7.4543e+00 -2.2254e+00 -1.9223e+00
#>  -5.2765e+00  1.1856e+01  1.2331e+00  1.3469e+01 -4.3080e+00  7.8455e+00
#>   2.0510e+00  6.2962e-01 -4.5395e-01 -1.7876e+00  6.4145e+00  2.5057e+00
#>  -3.3286e+00  1.1438e+01 -6.3853e+00 -1.8634e+00 -7.5483e-01  1.0974e+01
#>   4.5578e+00 -9.8753e+00  5.0932e+00 -4.8034e+00  6.6148e+00  1.1630e+01
#>   4.3757e+00 -3.7276e+00 -8.1431e-01 -2.1507e+00  9.3268e+00 -5.7862e+00
#>  -3.5055e+00 -7.3144e-01 -7.9797e+00 -1.7374e+00  2.4483e+00  1.3213e+00
#>  -1.5074e+00  6.0154e+00  5.0979e+00  5.1153e+00 -2.2462e+01 -6.2435e+00
#>   3.3304e+00 -1.5075e+00  4.9162e-02  4.6005e-01 -8.4947e+00  1.6674e+00
#>   1.5857e+00  1.9329e+00 -2.5193e+00  1.1608e+01  1.9823e+00 -1.3316e+01
#>  -1.5892e-01  4.6933e+00 -2.5752e+00  4.4518e+00  1.6552e+01 -1.1167e+01
#>   8.2605e+00 -7.8471e+00 -4.3856e+00  7.1784e+00  6.1026e+00  1.6421e+01
#>   1.2805e+00 -1.0181e+00 -3.8533e+00 -1.2263e+00 -8.9746e-01  4.1213e+00
#>  -1.6651e+00  5.2529e+00  1.0095e+01 -3.6652e+00 -3.7165e+00 -5.3011e+00
#>   4.5326e+00 -9.7920e+00 -8.6029e+00 -3.3408e+00  5.1334e+00 -1.6418e+01
#>   2.4926e+00  2.2211e+00  5.2480e+00 -8.7432e+00  2.6539e+00  9.0708e-01
#>   1.5093e-01  1.1564e+00 -1.5630e+00  7.2214e+00 -3.7623e+00  7.1154e+00
#>   1.3060e+00  5.4304e+00  1.2846e+01  1.1452e+00 -1.8506e+01 -6.4611e+00
#>   8.0697e-01  4.8919e-01  2.9823e+00 -1.1403e+01 -1.2208e+01  5.1964e+00
#>   1.8040e+00  1.2170e+01 -4.3492e+00  4.1412e+00 -1.6384e+01  1.2799e+01
#>   2.7019e+00 -4.0390e+00  4.9886e+00 -2.5779e+00 -9.3472e+00  8.0500e+00
#>   1.0112e+01 -6.3018e+00 -9.4359e-01 -1.1934e+01  1.6612e+01  7.8052e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]