Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.3499  -0.5268  -1.7227 -11.7233  -1.7600  12.5313 -12.7548  -1.2286
#>    0.2356   6.2994  -0.7748 -18.1438  -4.4464  19.8470   0.2408  -9.0953
#>    2.9842  -4.4653  -0.2673  -2.7484  -3.8885  -9.8691  -9.3695  10.8060
#>    4.3561   3.2717  -2.2271   7.7083  -6.3588 -20.6717  -5.6331 -12.2502
#>   -0.5784  -0.8016   5.0668  15.2813  -4.3691   0.5057  15.4678 -11.3727
#>   -2.4091   1.0178   7.8054 -13.9652   8.2771  15.0730   0.6972  -1.9485
#>    3.4425   4.5187   5.7795 -10.1346 -19.7304   5.1139 -13.8352   2.1907
#>   -2.6955  -7.7122 -14.2836   6.1267   4.7078 -14.5047   3.5200   5.8061
#>    1.9408   9.0328   6.7799  -0.3215   6.8008   6.3674  -1.2101  -0.0024
#>    1.9729  -5.6081 -17.6041  -6.7697   4.9608   0.9190   0.5629  10.9475
#>    3.4241   4.8464  10.3428  18.9647   8.5409  -5.5993 -11.8163 -16.0580
#>    3.2459  -1.1418  -4.2903   7.6169  -3.1815   2.7573   2.8833 -13.7148
#>   -0.3804  -8.2467  14.0069  -1.0063  -7.8801  -3.4130  -9.1865   7.7765
#>    1.9141   0.1903  -1.9498  -8.4174   8.6910   1.7236 -14.6388  10.5062
#>   -1.1865   0.5227  -3.8444  -0.2036 -14.2865  -1.5698   4.5503  12.8183
#>    2.1781  -1.7460   7.9910  -1.3074 -14.1135   6.6310   0.0855  -8.6042
#>   -1.4289   9.3216  -5.1191  12.0187   3.0978  -6.5634  -3.5721  -7.4849
#>    2.1770   1.7912   0.3543   8.6553  -2.0262   7.0579  -0.3444   3.1079
#>   -5.1554  -8.0200   4.3198   9.3389  -2.6403  -9.1086   2.9209  -7.0129
#>   -2.1775   6.6644   1.0375   8.3491   6.1376  -2.4624   8.9413   1.2517
#>    0.6229  -2.5190 -11.6803  -7.7390  -3.2633   7.4909   1.0029   9.8123
#>    2.7690  -1.1155  -3.5387   4.7047  -1.3666  -4.4472  10.5576  -5.7904
#>   -8.3823   2.7035   9.3168 -14.3157  -6.2132   8.5579  -3.9011 -11.6189
#>   -8.5049  -3.6535   5.5006  -0.0983 -14.8617  -8.3025   8.2566   3.0601
#>   -2.7495   2.7935 -12.0030   4.4939  11.7182 -10.1775  -2.8446 -12.2423
#>   -3.7533  -0.9581   1.5469  -2.1356  -1.6593   0.9138 -20.2489   7.6599
#>   -0.3768  -0.8009  -8.4676  -4.3213   4.1988   1.7660 -14.1904 -14.5432
#>   -1.0976  -0.7379   1.8565  11.5565  -4.6057  -5.8757   2.2844   7.1995
#>    1.7097   2.2222 -14.2457 -10.9373  -4.4784  -3.6547   5.1086  18.1014
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]