Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.7271  -1.9120  -2.1150   9.9091  13.2323  -1.6142 -12.3177  15.8833
#>    3.6261   7.1583   7.1510  -6.7849  24.6721   8.9320 -18.5694  -1.3361
#>   -3.9848   0.1365   7.3449  -1.5785   2.7702  -1.6330   8.5840   2.4186
#>   -5.6309  -0.2226   6.5328   3.1591  13.1313  -2.3962  -4.5005 -15.3670
#>   -6.3104   7.6284   4.7810  -2.9184   5.5766 -17.5822  -7.8715   3.7566
#>    1.0970   1.8392  20.0721   6.0554  -3.7244 -10.2719  17.1022  12.5459
#>    8.5861  -4.3863   7.2730  21.3477   8.5852   0.5686  -2.0140  -4.6212
#>    1.1701  -0.1849  -3.9395   9.4009 -28.5029  -0.4562   1.6817  -2.4162
#>    9.6832  -2.5862 -18.4965 -15.5527  -6.5128   6.5294   2.8144  10.4642
#>    1.2274  10.6468  -9.1734  -2.0241  -5.1165  -5.2989  10.9444   1.7158
#>    6.6854  -5.4943  -8.5258  12.6258  -5.8895  -1.9545  -9.5526  11.1005
#>    9.9776   0.8131   3.6988  -4.3100   8.6791   8.6203  -6.3665  -8.4684
#>   -4.0802   8.4694  -1.5780  -1.5911   5.6035 -10.3473   2.5067  13.3369
#>   -2.6563   7.8074  11.8607   9.3657   1.2962  -3.0025 -16.5466  -1.6301
#>   -0.3949   6.3951  -7.2223   5.8144   4.2774  -6.0238  24.9010   3.8562
#>   -8.3635 -12.1491   7.3976 -23.1427  -9.7042   6.3619   8.6096   6.4672
#>   -2.4317 -12.0149  -0.6363   2.8394   6.7263   4.5938   5.9225  -4.4092
#>    3.9816   0.4823  -2.4919   1.0720  -9.8993   6.2771  17.7401   4.4816
#>    6.8215   1.5972   2.4204   1.5681   5.9290   6.6489   5.2106  -8.2050
#>   -2.0431  -1.0350  -2.6757   9.0582   6.4962  -1.4566  10.9019   5.5231
#>    2.1555  -0.4960  -1.5381  12.4711  14.9843   3.2492   3.3126  21.6185
#>    5.7459   0.6347   0.4056 -10.2514  -1.5992  -5.8526  -2.2076   1.6526
#>   -5.1986   7.2347  -0.4262   0.4701  -5.7425 -10.5947 -12.9679  -3.7615
#>   -2.0469  -3.0661   8.5015  -2.7713   8.4067  17.4137  12.2439   6.7052
#>    0.5337  -0.7838   1.7793  14.5965  -1.8992   1.6496  -0.0226   1.1727
#>   10.2700   3.1906  -2.4521  -0.0248  12.2737 -11.4934   7.0047  -8.6634
#>   -3.2596   0.0642   3.3027   3.6438  -6.8346  -0.6054   7.2420  12.7339
#>   -8.7732   0.7815  -5.0316   7.3344  12.2081  13.8477   0.4228   8.0945
#>   -8.7921   1.2825  -6.3111   6.8027   9.3161  15.7162   1.0262  11.7973
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]