Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -4.7436e-03 -1.8196e+00  3.6896e+00  5.7809e+00  5.4551e-01  7.4523e+00
#>   4.1671e+00 -1.2730e+01 -1.5504e+01 -7.5826e-01 -2.6043e+00 -1.8464e+00
#>  -9.1329e+00  3.3540e+00 -1.0062e+00 -1.2710e+01 -1.0822e+01 -1.4343e+01
#>   3.1100e-01 -1.0072e+01  6.5770e+00 -4.7766e+00 -1.8472e+01 -1.5269e+01
#>  -2.0336e+00  7.3568e+00 -3.3516e+00 -4.9581e+00  3.6997e+00  5.6020e+00
#>  -1.8018e+00 -1.0935e+01 -3.1225e+00  3.9084e+00 -4.1479e+00 -5.2163e+00
#>  -6.0866e+00 -2.3792e+00  1.0900e+01 -5.4972e+00  2.4875e-01  3.8394e+00
#>  -2.0612e+00  5.9615e+00  6.6517e+00  7.0064e+00 -2.7528e+00  2.8628e+00
#>   3.0590e+00  8.0632e+00 -4.2852e+00  1.7020e+01 -3.1705e+00  9.4582e+00
#>  -7.7886e+00 -5.5593e+00 -2.2999e+00  3.3426e+00 -5.1428e+00 -1.6436e-01
#>   2.3667e-01  4.2221e+00 -8.3020e+00 -5.5349e+00 -5.9952e+00 -1.9321e+01
#>   3.3463e+00 -8.4463e+00 -3.0156e-01 -6.2206e+00 -1.0708e+01  2.9310e+00
#>   3.5630e-01  6.3200e+00 -8.7539e+00  5.9655e+00  5.9884e+00  6.8972e+00
#>   3.4906e+00 -1.1421e+01 -6.1892e+00 -2.2423e+00  2.1750e+00  5.3952e+00
#>   4.7463e-02 -2.0062e-01  2.0440e-01  9.6992e+00 -3.9032e+00 -2.1292e+00
#>  -9.3323e-01 -1.2665e+00 -9.2105e+00  6.6602e+00 -7.4648e-01  2.6415e+00
#>   2.3078e+00  7.3406e-01 -2.2773e+00  1.0659e+00  2.6763e+00 -5.9007e+00
#>   6.0702e+00  1.5031e+01 -4.2910e+00 -2.0649e+00  1.2161e+01 -7.9364e+00
#>   1.4768e+00 -6.9294e+00  7.5518e+00  3.1378e+00  7.4308e+00 -3.2485e+00
#>   5.1526e+00  7.0187e+00 -8.5282e+00  3.5646e+00  5.5774e-01  5.7214e+00
#>  -3.8560e+00  5.8338e-02  8.9972e-01 -9.0925e+00 -2.2760e+00 -1.8721e+00
#>  -1.3751e+00 -1.9334e+00  9.9197e-01 -1.0426e+01 -1.5369e+00 -1.4546e+01
#>  -5.9174e+00 -1.1679e+01  6.3497e+00 -1.7750e+00 -2.1739e+00 -1.3953e+00
#>   3.1717e+00 -3.1103e+00  5.4266e+00  7.3469e+00 -5.2673e-02 -1.4664e+01
#>   1.8356e-01 -6.3521e+00  5.5915e+00  4.6971e+00 -1.1767e+01 -1.4247e+01
#>   5.8762e+00 -3.2301e+00 -1.1271e+00 -1.1769e+01 -3.1609e+00 -4.0091e+00
#>   8.6240e+00 -6.7850e+00 -7.0539e+00  8.1127e-02 -3.5996e+00  3.0483e+00
#>   2.6945e+00 -5.9041e+00 -3.7268e+00  7.8381e+00 -7.7786e+00 -1.1700e+01
#>   5.3412e-01  3.9201e+00 -4.8311e+00 -1.5003e+01 -8.5240e-01  1.1715e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]