Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -4.4771e+00  1.0809e+01 -7.8645e+00 -1.2694e+01  7.5395e+00 -3.4418e-01
#>  -1.4609e+00  9.8784e-01 -6.4002e+00  4.9026e-01  7.8326e-01  2.2227e+01
#>   7.7685e+00 -9.6896e+00  1.2211e+00 -6.1264e-01 -5.4945e+00  3.8134e+00
#>   5.4466e+00  1.6241e+00 -3.3300e+00  3.4401e+00  1.8142e+01 -2.7119e+00
#>  -5.0096e+00  1.7162e+00 -3.5789e+00 -2.0780e+00  4.0965e+00  3.1284e+00
#>   7.0256e-01  1.4636e+01 -1.3289e+01  4.3312e+00 -6.3862e+00 -7.8085e+00
#>  -1.2547e+00  5.1936e+00  5.9812e+00  2.3031e+00  5.2983e+00  1.7876e+00
#>   1.9928e+00  3.3712e+00  1.5151e+01 -7.3841e+00 -2.2570e+00  1.2005e+01
#>   4.1630e+00 -7.9772e+00  1.1355e+01  5.1849e+00 -1.5179e+01  1.4419e+01
#>   2.8217e+00  6.7798e-01 -6.1731e-01 -4.2816e+00  1.3266e+01 -5.8433e+00
#>   6.7763e-01 -8.5197e+00 -7.0548e-04 -5.8896e+00 -1.2367e+01 -1.3994e+01
#>   1.2257e+00 -5.1571e+00  1.2033e+00  5.9974e+00  2.3801e+00  5.0185e+00
#>  -4.4314e+00  3.8668e+00  4.1813e+00 -5.3909e+00  2.4883e+00  1.7067e+00
#>   4.7100e+00  4.0188e+00  3.3329e+00  1.8302e+00  8.8622e+00 -8.4238e-01
#>  -7.1020e-01 -8.2925e+00 -4.6394e+00 -9.7713e+00 -4.4881e+00  1.1136e+01
#>   1.0914e+00  3.0081e+00  3.3788e+00 -1.0462e+01 -1.2539e+01 -4.5281e+00
#>  -5.5134e+00 -2.6218e+00  1.1016e+01  8.8876e+00  2.8332e+00 -6.7477e+00
#>  -5.5394e+00 -9.1376e+00 -6.4610e+00 -9.3604e+00 -9.5914e+00  5.8475e+00
#>   1.4510e-02 -3.5538e-01 -3.5695e-01 -8.6142e-01 -8.3755e+00  7.3272e+00
#>  -3.2375e+00 -2.2184e+00 -1.9192e-01  2.1490e+00 -1.1073e+01  5.1968e+00
#>  -3.7242e+00 -1.0482e+00 -8.6973e+00  1.8418e+01 -9.3359e+00  1.5658e+01
#>   1.8085e+00 -7.3037e-01 -4.9894e+00  2.4340e+00 -1.0740e+01  3.3479e+00
#>   1.0974e+00 -4.1871e+00  1.0304e+01 -1.2982e+01  1.0600e+01  1.3708e+01
#>  -1.2241e+00  1.6004e+01 -1.7056e+01  2.0829e+01  4.0100e+00  9.4019e+00
#>   7.1123e+00 -2.5625e+00 -9.2219e+00 -1.8388e+01  1.9445e+00 -3.2628e+00
#>  -3.2123e+00 -5.9031e+00 -1.6828e+00 -4.6543e+00 -8.1021e-01 -4.2286e+00
#>   2.8816e-01  5.6977e+00  1.0034e+01 -1.4337e+01  9.8545e+00 -1.0529e+01
#>  -4.3377e+00 -4.3917e+00 -6.2792e+00  3.0741e+00 -6.2360e+00 -3.7564e+00
#>   1.8151e+00 -8.3609e+00 -6.3080e+00 -6.6703e-01 -9.6076e+00  5.2527e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]