Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.5629  -8.9843   0.8461   4.5530  -8.1209   7.2343  -3.0359  -3.3137
#>    0.3056  -6.8177  -8.0021   6.3591  -0.9077  14.6663  -8.6466  -9.6905
#>    4.5901   2.2485  -7.3138  -5.9730   1.9744   5.0015   8.5098   3.2875
#>   -7.5171  -8.5900   9.2048   0.4057   0.6250 -15.9217  -8.0800  -0.6859
#>    0.9597   4.0740  -1.6640   5.4757   6.3296  -7.7157   2.8335   2.3738
#>   -0.3562  -5.2483   9.4585  12.1767 -12.5496   4.9008   0.7798  15.4638
#>    3.9203   1.2213  -5.1311 -12.0941  -1.0798   3.8762   2.8602  10.8809
#>    0.0601   6.2650   1.5262   1.7433   1.6945  -8.5649  -3.3902   6.1645
#>    2.0105   3.4180  -1.7960   1.0258  -1.5575  -2.8825   3.4014  16.8274
#>    3.0581  -1.2408  -7.3476   9.2582  -7.5854   1.5390  -6.2489   5.2142
#>   -2.1199  -0.8779   1.5146  -3.2521   2.9631   9.8150   3.9508 -14.1756
#>   -2.2799   7.0862  -6.2405   2.8846   1.2971  15.2180  -0.9488   4.0208
#>    0.9591   2.4601  -3.4362  -6.8922  -6.0736  -6.9197   6.6142  18.2761
#>   -4.0475  10.3972   4.0440   5.4563  -4.5331   9.7272   3.1252 -24.6474
#>    0.5154  -3.9788  -2.3733   6.8449  -2.0285  -3.2062   5.4585 -11.0804
#>    4.4616   2.5741 -10.2579   3.8001   2.3770  -5.3347   9.4679  -1.3492
#>   -3.9385  -0.7719  -0.2824   8.2519   1.7579  17.3771  -9.4187  -7.4151
#>   -1.7539  -4.3432   2.0333  -5.0899   5.2515 -15.4635   1.6629   1.7947
#>    2.5663   2.2303   9.5988  -2.9217 -16.0758   6.5208  13.7187  -1.9241
#>   -2.3497   5.0226  -8.3884   3.0610 -10.7386  11.0149  10.5128  -0.4737
#>    3.7330  -3.6045  -2.1795  -6.6258  -9.3199   0.7699  -5.3556   6.4614
#>   -0.3836   0.7955  -1.6729  -4.4979   6.8379  -2.5984   6.5493  -6.0061
#>   -3.6087  -3.8580   6.2626   0.6361   1.2442   7.2649 -24.8001   2.0922
#>   -2.0990   0.4330   9.4291   3.2798  12.2606   2.8194   7.5711  -9.3627
#>    0.8064   1.5159  -6.2875  -3.1592  -0.1626  -3.4103   2.9500 -15.5269
#>    0.3293  -0.4159   6.2014  -4.0438 -13.4347  11.4114 -17.0941   6.1478
#>    3.3024  -4.9779  -6.0239  -3.1668  -5.6188  -7.4908  -1.4566   9.1998
#>    1.7148   3.3087  -1.6150   3.5956   0.8701  18.9874   5.3405  -0.5779
#>    4.5314   7.2358  -2.4302 -13.2462   7.1177   7.8388  -2.1773   7.3880
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]