Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.5428  -1.4827   2.2600  -2.4582   9.3076  -5.5957   3.3819   8.6445
#>    1.6904   6.6309   2.1330 -13.7717   6.5676 -13.1643   8.2698  -7.4368
#>   -0.3898   5.6509  -7.1689 -13.2007 -18.2438  -1.3541   0.9436  -0.8365
#>  -12.1324   0.0149   4.0324  -3.4427  -5.4998   7.2697  -2.9637  10.0165
#>   -2.0082   0.9718   2.9933   8.7925   1.1132   0.5340   3.2511  22.1300
#>   -4.4689   5.9727   8.8492   4.4854   4.8938  -0.8070  -7.4855   0.7584
#>    5.2900   7.5839 -13.7783  -9.7290   5.4130   5.4295 -10.2580 -17.6647
#>    3.7072  -3.9714   4.6667   5.6871 -13.2559   4.5061   1.4706  -3.5588
#>    5.7612   6.1874   9.3104   0.5659   5.6682  21.4556  -4.5171  -3.3186
#>    2.2322 -12.1055   6.5679   5.8309  -4.5961  12.5542  14.1560 -14.6685
#>    3.9165  -3.7364   0.7787   5.5085  -8.1523   4.0058  -6.2095  -1.5094
#>   -4.2002 -12.6036  -3.1682   3.3901  -3.5375  -2.9805   6.4784  -6.8285
#>   -0.9672  -1.3116  -4.5570   2.6595  -1.1091  -3.7821   9.7980   3.0526
#>   -6.0223   3.1025  -8.7741  -1.5638  -1.7258  -2.0704   8.2196   6.8509
#>   -4.2492  -4.8595   7.6374   7.6418 -12.1324  14.5408  -3.2015  -4.0341
#>    5.3214   1.3502  -2.5592   5.4787 -12.8718   1.2674   7.0179   1.3328
#>    5.1562   0.3929   9.2172  10.7752   2.1644  14.9945   2.9473   7.1872
#>   -1.3364   3.4805  -1.5276  10.7889  -1.2891   2.6430   0.5381   3.1557
#>    7.0689   7.1165  13.9607   9.4669   1.2699  -3.4414   4.7536   1.6705
#>    3.8127  -7.4588  -4.3932  10.6924   6.5721  -3.0043  -2.2741 -12.5714
#>   -3.3132   2.6595   1.2683  -5.9301  13.3356 -18.3298   7.9474   9.6479
#>    4.4433  -7.1111   6.7120  12.9453  -3.3309  -9.1969  10.2559   6.7810
#>    5.0510  -6.9601  -9.1161   5.1178  -1.9025 -12.6464   0.9989   1.4527
#>    2.9625  -3.4121   5.6125   6.5992  -0.8156 -15.5830  19.7037  -2.4046
#>    1.2153  -0.6656   1.5626   0.6113   1.2127   4.1468  15.9135   4.6278
#>    5.6581   5.4200  -3.0155  -5.8477   3.0419 -16.1614   6.6569  11.0573
#>    1.5731   5.8084   4.7874   4.1126   8.1719   2.8335  17.5286  24.9425
#>   -4.2448   1.5183  -4.3200   2.5963  -5.6719  -2.9386   0.2236 -13.2904
#>    4.1144   3.3727  -0.8298  -1.1120   4.2595  12.9041  11.1415  13.7667
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]