Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  7.2623e+00 -1.7055e+01 -8.1537e+00 -6.3030e+00  2.2526e+01 -9.2313e+00
#>   1.2945e+00 -4.6094e+00 -6.6874e-01 -6.2709e+00 -8.0662e+00 -9.8570e+00
#>   3.6372e-03  2.7183e-01  1.4382e+01  8.9109e+00 -5.4207e+00  5.8586e+00
#>  -3.1104e+00  7.4930e+00  9.1830e+00 -6.7250e+00 -9.7572e+00 -2.6488e-01
#>   1.2463e+00  7.5699e+00  1.5752e+01 -1.7233e+01 -1.0988e+01  6.5096e+00
#>  -6.0034e+00 -1.1895e+01  1.5994e+01 -2.5725e-01  2.0101e+01 -2.0946e+00
#>   9.4055e-02  9.3410e+00 -8.1753e+00 -4.3880e+00 -1.1199e+01 -1.1815e+00
#>   8.0030e-01 -2.1163e+00 -1.0413e+00 -1.0789e-01 -1.4620e+00  3.5262e+00
#>   8.5993e-01  8.8738e+00  4.6152e+00  5.8162e-01  8.1022e+00  2.9554e+00
#>  -5.5466e+00 -9.2410e+00 -2.7887e+00  4.0144e-01  1.3152e+01  8.9877e+00
#>   8.7762e+00  5.3933e+00  1.8271e+00  4.8730e+00  5.4257e+00 -9.7746e+00
#>   6.0982e-01  2.3967e-02  5.3979e+00 -5.8992e+00  1.8067e+01  1.5937e+00
#>   5.2943e+00 -2.6622e+00  6.0488e+00 -1.1476e+01 -2.6782e+01 -2.1627e+01
#>  -5.6996e+00 -1.5072e+00 -2.8376e+00  1.2578e+01  2.6264e+00 -2.2501e+00
#>   2.2144e+00 -8.0781e-02 -1.3545e+00  5.7635e+00 -1.4623e+01 -1.9362e+01
#>   7.1008e-02  1.3292e+00 -3.0182e+00  1.1991e+00 -4.4152e+00 -3.6085e+00
#>  -4.2953e+00  6.8780e-01 -9.2720e+00  5.8672e+00  1.0984e+01  7.7386e+00
#>   3.4999e+00  9.5388e-01 -5.5494e+00  5.7161e+00 -1.1978e-01 -3.3832e+00
#>  -6.9111e-01 -5.1461e+00  1.0320e+01  2.8589e+00 -2.2605e+00  1.1588e+01
#>   4.4698e+00  5.3238e+00  4.4958e-01 -9.7556e+00 -9.0811e+00  4.6425e+00
#>   2.9048e+00 -1.1800e+01  6.5441e+00 -4.0105e-01 -7.2522e+00  7.0799e+00
#>  -9.5608e+00  5.6030e+00  2.9562e+00 -3.4275e+00 -1.5356e+00  4.3073e+00
#>  -2.7037e+00 -1.7313e-02 -1.6407e+00 -7.2330e+00  1.5195e+01  2.5670e+00
#>   2.7652e+00 -4.4173e+00 -8.3351e+00  2.8537e+00 -4.7832e+00 -5.4605e+00
#>   3.8609e+00  8.5058e+00 -3.4572e-01 -8.6110e+00  1.4051e+01 -7.3177e+00
#>  -5.0703e+00 -6.0179e+00 -2.7001e+00 -9.1383e+00  1.1483e+01  1.2736e+01
#>   3.2076e+00 -5.3511e+00 -6.4300e+00  5.7287e+00 -3.9354e+00 -2.9230e+01
#>   1.6252e+00 -1.1506e+01 -1.3173e+00 -1.4891e+01  1.8824e+00  1.1005e+01
#>   3.2842e+00 -2.5989e+00  6.7036e+00 -2.3197e+00  1.5627e+00 -1.6391e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]