Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8 -13.8272  -9.8711   1.3734  -2.3926   7.6634   8.4909  -1.6599   9.0782
#>   -2.2394  -9.5204  -6.1462   8.7949   8.0176   2.9166  -6.3172   4.7824
#>    4.3900  -5.0629 -10.4900   7.0041  12.0077 -15.4577   6.8375   6.6954
#>    9.7755   8.4922  -5.7388  10.5360  -3.3402  -0.3472 -10.3201   1.7089
#>    2.4098  -0.1047  10.4052   0.1811 -12.4194 -15.6430  10.0909 -13.7841
#>    0.7829  -1.8784   1.7206 -10.2101   4.6157   9.7850   1.0750  12.1911
#>   -3.9546  11.5559   9.5122 -11.4937  12.0279   7.5032   4.6365  10.8678
#>    2.2263  13.5508  -4.1743 -11.8053   5.8218   5.0347   3.8473   0.5270
#>   -7.4935  -0.3251   7.8679  -0.8273  15.4264  -3.8030  -9.9392  14.7772
#>   -9.2605  -6.2653  16.8711  16.4731  15.3423  -1.7635  -7.6836  -9.4011
#>    0.6269 -15.0027   6.8913  -1.8266   6.9700   7.8065  -5.0864   1.6773
#>    9.5546   1.1979   2.4524  12.6340  10.8644   1.9642 -18.2777   1.3054
#>  -13.4933 -14.9476  16.3241  10.0407   3.4798  -4.7022  -9.1301  -5.2573
#>  -12.7988 -10.4053  -5.2676  12.9979   9.1790  -9.8193   3.1260  -5.6041
#>    0.9062 -10.1379 -10.5995  -1.6110 -12.5792  -4.7032   2.7020  16.6105
#>   -4.4958  -2.1508  -6.6083   0.0497   2.9082   5.4202   0.3307  11.6487
#>    6.9060   7.1612 -15.5791  -2.5193  11.2413  -4.5306  10.8163  21.2856
#>   -0.9794   1.5260  -8.9173  -3.1960  -2.5324  -0.5590  -6.9675  10.6801
#>    0.9619   3.1111  -2.8311  -8.6374  -5.4336 -15.4684   9.7290 -12.4803
#>    0.8941   3.1967  10.7350  -3.9782 -14.0671  -3.8283 -10.3961  -3.0575
#>    6.2559   9.5132  -9.7572  -1.1027 -15.9669   1.7399  -7.1200 -16.7836
#>    2.1720  -3.7063  -3.9736   4.3831  -5.6018  -2.8353   5.4292   3.7560
#>    1.9453   5.0000   2.8417  -8.2360 -10.7713   6.5683   2.9042 -13.2643
#>   -8.9009   7.6448   6.3579   5.7123   6.1545 -13.1776   7.1872   7.2590
#>   -9.2325  -5.7587  -1.8703  -4.2580  -0.7857  -8.2359  -1.8056  -0.0224
#>    0.8857  -7.5567  10.6017  -6.1956  -9.4431   1.9686   7.5976  -9.6724
#>   -1.5276   0.8935   6.3486  -2.7168   2.5480   7.9043  -3.2904   2.7325
#>   -1.4520   7.0022   0.9542  -0.0847  -1.9477 -13.0276 -15.1538  -5.8920
#>    0.3973   4.4487   5.8710   1.4361   6.4730   7.7392  -0.8810 -10.9044
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]