Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.6850  -1.0182  -4.1355  -6.3000  -1.8725   0.6188  -8.5158  -1.6497
#>   -1.6162  14.7757   3.8566   6.1701  -5.9799 -31.9878  -4.6860 -17.4054
#>   -3.4152   6.3494   7.3595  -4.8106   9.0418   2.5725  -1.3882  12.4305
#>    2.1526  -4.8367  10.8095  -7.1143   0.5264  -5.0508  -4.3906 -11.2740
#>   -3.0232  -3.7416  -1.2734  19.7394   3.3349  -5.0737   7.9322  -6.1717
#>    0.5244  -2.3895   3.3659   4.6299   3.4481  17.8751  21.3284   0.5163
#>   -4.6555   8.6851   0.7325   6.1966 -10.9072 -19.3979 -15.4315  -6.2822
#>   -2.3843   2.0748   2.8696  -1.6785  -9.1864   0.3065   2.6150  -2.6091
#>   -3.5424   1.6787  -6.1689 -13.5930  -3.4280 -19.8133  19.2829  -8.7828
#>    3.1923   4.6201   2.1399   6.2290  -2.1391   6.3301   2.6819  -0.8586
#>   -0.9861   0.2242  -9.2001  15.0113   3.9226  -0.0312  -6.7753  -5.2870
#>    1.4526   7.2566  -3.1013   7.4679  10.8411  20.6621  15.9681  -5.3786
#>   -2.0635   2.1675   4.5908 -13.6509  13.6747  10.9121  14.7220 -23.6261
#>   -3.7459  11.0049   0.0157 -17.0377 -13.8138  -0.0395   5.4398  10.7543
#>   -2.3084   4.0497   0.6070  -4.3288  -5.8547   0.0094  -1.8816 -12.3234
#>   -2.6460   8.0173   1.3882   9.0476  -1.7375   7.1406  -0.5467 -21.9535
#>    3.8697  -7.7212  -4.1873  14.1685   1.5498 -11.6353   2.3459  -1.7476
#>   -0.2430  -5.4848  -5.0722 -19.0946  10.1127 -11.1649   6.8031  -8.3274
#>   -6.7372  -0.1400  -2.0475  -5.3410   8.1227  14.6745   5.4684   0.1166
#>    4.0562  -3.7951  -2.4679   1.5395   4.6035   7.9561  15.2640   3.9575
#>   -3.0955   6.8233  -6.0318  -2.7500   3.0807  -2.1661   9.6980 -13.7891
#>    3.5770   1.4862   2.9603   3.4413   8.4962 -12.7599   9.6074   6.3584
#>   -0.0835  -0.1856   0.9578   2.2782 -14.3954  -0.6548   1.9030  -8.5268
#>   -2.3208  -7.0679  -5.7317   2.3746  10.7740  -4.5407  -0.6261  -5.1855
#>   -0.1899   1.6089   0.5425  -5.2691  -3.3635  -4.9586 -21.7374  -7.6439
#>   -5.1970   7.4059 -10.1111 -17.1562   9.2674  -1.6337  -8.1260  18.3789
#>   -0.1408   0.8293  -4.4535  -5.1899  -8.5529  -3.2217   4.2348   1.6676
#>   -0.9824  -4.2816   3.0957   2.5420  -4.8419  -3.8307 -21.2491   0.4303
#>   -2.3623   0.5354  -9.4953  14.8029 -19.1285  23.9714 -12.5011  -0.5608
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]