Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.8308  -6.3955   5.2306  -3.2833  -6.6411  -0.3573  -8.0694  -2.9305
#>    4.3831  -5.5681   1.7187  12.8080   4.7056  -4.7151  18.0986  -4.4075
#>    8.4424   3.1957   1.5913 -12.5117   3.4321   3.3238  12.7277 -10.4031
#>   -0.4442  -5.1953   9.7377 -11.4245  -3.8977   1.6733   0.2445  -6.0568
#>   13.1379  14.0881 -13.6817  13.7449  12.2518   7.2010  19.0655   0.5370
#>   -5.7658   1.1302   6.4924   4.8661  13.2448  23.6401 -14.4389   6.6485
#>   -5.9113   1.7592  12.6699   2.1303   8.7393   2.2839  -4.1279  -1.9399
#>    1.2884   7.4567   2.7563   0.3666  -7.5620   3.4011   6.0272   0.5101
#>    2.2089  -7.6015  -0.5015  -2.7806 -16.0505  12.5940   9.4860  -7.4753
#>   -0.6220   9.4047  -3.1724  17.3560  13.7204   5.4274 -13.3513  14.2916
#>   -1.3035  -3.5716  22.1038   3.7513 -14.4116  -5.6337  -7.8527  -6.0463
#>    9.7960   2.9040   1.0416  -5.9612  16.6282  -9.3591  -4.6564  -1.2448
#>   -3.8701  -5.2524   7.4280   2.5846  -9.4459  -8.2586   9.6812   5.3286
#>    2.4764  -0.5858  -7.8516   0.3647   2.4530  29.6826   8.6991  -1.1407
#>    9.2938   5.9732  -2.6255   2.7612  -5.5303   5.8777   6.7586  10.2884
#>   -1.7310   9.5054 -18.1512  -9.1872  11.6857  11.1371  -1.5081   1.2402
#>    3.3571  -5.4254  -1.8523   5.2894  -5.4338  -7.5513   7.6492   5.0221
#>    3.1889   7.2298  -9.5049  11.4072   8.3108  -1.9175   7.3308  -4.1825
#>   -4.2742 -16.5679   0.9181  15.9307  14.2398  14.0169 -11.9570  -7.7918
#>   -2.6042  11.8571  -7.8159   7.2703  -9.1024  12.3101  -4.9255  -0.1127
#>    8.1727  -9.2728  -3.7509 -11.4498  19.5286  14.0060  -1.5843 -12.0478
#>   11.2030  -9.9668 -15.3645  -3.1676   0.4477   0.8577  14.4209   5.9550
#>   -4.8428  -3.5444  13.7712  -3.9883  -3.6272  -4.6103 -11.1632   7.5163
#>   -2.6318  -1.8517  -5.9832  -4.2043  -8.6386   8.9081 -11.6017   5.5028
#>   -7.9695   1.1148   5.8575 -19.8438  -8.8753   8.2195 -13.8496  -1.7425
#>   -7.1294  -0.8212   1.7956   5.1679 -10.9385  -4.0638   2.2980   4.7989
#>   18.8447  -1.8777   3.0676 -10.6042  -3.9478  15.0084   0.9576   9.6767
#>    1.8429   3.4556  -5.8065   1.0721  -6.5602  17.0868  -3.2106  -3.8284
#>   -2.3781  -0.9807  -4.3341  -3.0685  -7.9337  -1.1491  13.3449  -6.9589
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]