Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.7620   1.5723  -5.8995   8.2917 -13.3628   2.2746   0.9772  -2.5725
#>   -4.8759  -0.9411 -13.6213  -2.2590  36.7604   9.5877  23.8199  -4.4564
#>    1.5425  -5.5736  -9.5090  -8.8519  10.8315   0.5849  14.2184  -1.4516
#>   -0.6487   1.5594  -4.2069   6.7883   2.3546 -31.7186   0.9366   7.8962
#>    2.1211   7.0699  -4.0567   1.9084   0.4989   5.0114   0.3853  13.5821
#>    6.7115   1.4928  -0.5338 -15.7068 -17.0926  11.0502   6.0667   5.6939
#>    2.1807  -2.2815 -10.4594   4.5475   4.2241   9.9215   6.0245   7.3310
#>   -2.2679  -0.8183  -0.2417  13.2124  -3.8465   6.2797  12.3444   1.0951
#>   -4.5520   7.5301   8.5288 -11.0302  -2.3486   2.3179  12.0568 -16.4425
#>    1.6022  -0.9674   2.3702   1.2989  -1.5135   2.3499  -4.2604 -10.8292
#>    7.3201   1.3586   1.7084 -10.2937   3.6417  -0.9937   6.2683   8.2310
#>    0.2922   5.4732  -4.6838  15.8207   2.8045 -18.4356 -23.8940  -6.4264
#>   -1.5474  -1.7215  -4.7036   1.8585  -5.7962   8.3928 -12.5442   7.8024
#>   -0.0919   2.2665  11.7521  -8.3422 -11.8980   5.7438   9.2175   5.2106
#>   -4.2248  -4.1427 -14.2900  10.1195 -12.6154  -1.9329 -12.0703   5.7775
#>    9.1677  -4.2215   6.8736 -14.5285  -6.5751  -3.0086   1.9398   2.8348
#>   -8.0089   3.8284  -1.9211  21.4771  12.2642  -5.7115 -23.2092  -1.9791
#>   -7.9165  16.0474   0.7708   5.3623   1.4415 -15.8901   4.8257 -17.1509
#>   11.9495 -11.9859 -10.3559  -0.4229   4.8186   1.7919 -15.6334  21.4093
#>    5.0902   0.0800  -3.6769  -3.3695  -2.3464  -3.9127  -8.6190   3.9887
#>    1.6607   0.4218   3.0811   3.9800  -0.8410   3.2879  -2.1618  -7.5608
#>    0.9997  -6.2647  -3.9647   2.3731  13.5472  12.3726  -1.5128  -0.8107
#>    2.2679  -3.1995   8.6364  -4.7186  -3.4246  -2.8491  19.6437  10.2708
#>    8.4817  -4.8458   1.7444 -13.7857   3.2218   9.3100   1.2856  -9.3520
#>   -1.5454  -4.4498  14.5576  14.7776 -18.8714  -9.7616  -3.4803  18.4943
#>   -3.6102   5.3527  10.0148  -2.5333   1.2581   1.5620   3.3339   2.3117
#>    1.4435   1.9494  -4.0764   0.8517   9.9613   1.5624  14.1459  15.0352
#>   -2.2282  -1.7571 -10.9574   8.3164 -26.5484  12.1686  10.7814   7.3996
#>   -5.9264   3.6918   6.2777   4.3380  17.6654 -15.0931   7.0969   4.1319
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]