Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.0024  -7.7467   1.0957 -24.3145  -1.3755  -4.6542 -11.6984   7.2271
#>    5.7724  -3.2683  14.0793  11.4518  -3.7649  -5.0841  13.4240   7.4938
#>    5.0456   1.9223 -13.1530 -19.6496  17.6345  -7.2471   8.7294   5.3340
#>    0.1954  -4.5144   9.4442   9.2103   2.0559  -1.3329  -1.6365   6.1389
#>    5.8160  -3.7487   0.3595  -2.5100   6.7922 -13.6264  -1.7207 -12.5043
#>   -1.4251 -10.0153   6.5411  -8.1921  -7.9266  -1.3991   3.0689   8.8792
#>    2.2772   2.2219  17.2605   7.7282  17.8373   7.7812   5.4229  -3.8948
#>    3.0088  -2.0478   3.5397 -10.4960   1.6709   3.0074   0.9324   3.5334
#>    6.7740  -2.0980  11.0897   4.5094   6.9446  13.6756   5.6351 -11.5937
#>    1.0248  -0.2678  -3.8326   9.1671   7.2737  -4.4055  -9.7316  -6.4788
#>   -5.7889   0.8520  -3.1936  16.3144   1.1963   6.4413  -8.6951 -15.7213
#>   -0.6060  -0.7139  -8.4866  -9.1876  11.9832  -2.2754  -4.2966   0.4287
#>    3.0038  -3.4785  18.5154   4.2578  -6.8311  -1.1053   6.3590  17.9475
#>   -2.6600   3.6664   7.8284  -5.2566   3.8084  15.7084   5.6277   3.6743
#>   -1.8058  -2.4919   6.6899   0.5308  13.2115   8.1422  -2.4952  -0.9409
#>   -4.6346   6.0589  -0.8645 -10.1172   5.4603 -19.6995   2.8750  -3.4651
#>   -5.9133  -4.1456  -5.6421   6.9561   4.2225   2.8033   7.6889 -11.3118
#>    2.1053  -5.6047  -2.4189  -3.4386  -8.2866   1.5663  -6.8965  -7.6138
#>   -3.5048  -1.2113  -4.5703   7.4817   2.9161  -4.5869  -8.2014   2.5774
#>   -2.9320   4.9896  -2.5636  -2.7155 -12.9661  -5.7788  -1.1553   4.1457
#>    1.0900   7.0868 -11.5598   7.5304   0.8783   6.0689   4.4494   2.4680
#>    4.3105  -5.4812  -0.0055   6.7588   5.0692   5.8935 -17.8311 -18.4819
#>   -7.6906   1.4343  12.9048  -2.4434   3.7112   2.4145   6.9602  10.7760
#>   -3.5377  -7.3421  -3.1822   3.0415   4.3981  -1.3497  15.1326   4.7561
#>   -1.5352  -8.0714  -1.8140  11.0685  18.3995   3.8210  12.3880 -15.8022
#>    8.5759  -1.7379  -6.3782   1.4072  -7.4659   7.1268  -1.3992  13.9952
#>   -4.3948  10.0308  -7.8816  -4.4612  -7.4225   1.3345  -0.0185   7.4674
#>    1.0118   0.9024  -2.6258  -1.5087  -1.2078  -9.8515   3.7993   3.9703
#>    4.6661  -8.0247   4.5591   5.1206 -13.6448   3.7407   5.3368   8.0352
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]