Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.2540  11.8975  -3.7024  11.5528  14.1356   3.6873   9.2732  -4.4685
#>    0.1873   7.1043  -3.5375   6.6741   4.7690  -1.6354  -7.8192  -8.1437
#>   -0.1012   0.2073   6.4304  -6.0280 -10.4169  13.2261  -0.3910   4.3322
#>    2.6035  10.4910   7.3667   1.1905 -10.8071  -4.7460  -3.5816   8.8186
#>    6.6931   1.2424  -0.1255  -3.5543   5.8229 -11.3437   6.5476  -1.7374
#>    0.6340   0.4253   0.2655  -0.8579  -0.2429  -8.3723  -2.1939  -0.9267
#>    3.9713  -2.0838  -5.8428  -0.6061   7.3152   1.7703  -9.9976  -2.0901
#>    0.2045  -2.9981  -0.1022  -7.4338   7.0959   9.4172   3.1132  -5.7134
#>   -2.1384  -8.6619  -8.1999   9.9916   4.4676   5.2301  -7.0237  -3.1638
#>   -2.5652   0.5026   0.8527   1.6994  -2.3280   2.0593   0.8557  -0.4984
#>   -3.4138   4.0256  -0.2309  10.7089  -3.0633   5.5823   2.1218 -11.5748
#>   -4.2133  -2.1193   9.9053   0.4915   5.9488   6.0590   2.0120   1.7995
#>   -2.6067  -2.8725  -5.2875  -3.1813  11.2466   3.1617  -2.5937  13.0268
#>    3.0496   1.1321  -3.4715  -8.2281   4.8672   4.9765  -0.0497  -0.3970
#>   -0.8846   1.5797   2.3607   7.9506  -1.2853  -7.6428   4.6481  -7.3056
#>    2.9885  -0.4625  -4.0789  -2.9286   1.1966   6.4835  11.6528   3.0655
#>   -1.5255   4.4864   7.6608  -2.0332 -16.3837   2.4966   7.7924   6.1536
#>    0.0906   1.5388  -4.3328   1.8019 -13.8467   9.6267   2.9559   9.6474
#>    4.7243  -2.1443  -6.1621   4.1490  -7.7168  -6.2285  23.5670 -18.2967
#>  -10.0733  -1.8383   4.1561  -0.8908  12.9019  -7.8946  -4.6562   5.5012
#>    1.7709   5.4345   8.1476   1.3488   0.8453   1.3269  -2.5381  -2.5422
#>   -2.7201  -5.9397  -5.1052  19.4429   2.4422   7.6962  -2.5661 -18.1620
#>    2.9294   1.3180  -3.2222  11.0130  20.0450  -1.8550  -4.1055  -1.9695
#>    2.3588  -4.6349   0.9215  -7.8586  -6.4369  16.4108 -13.1895  15.9012
#>    3.3786   5.7980   1.9379   9.8442  -4.8074 -10.6415  -2.3339  19.2941
#>   -0.7231  -1.6565  -3.1843  -8.4201  -4.0986  -4.2693   3.1281  -3.9198
#>    1.7196   1.9156  -6.6680   6.3365   6.2965   4.2454 -11.8782  10.3615
#>    3.9278   4.2588 -10.6050  -1.3672  -1.5539  -0.4058   9.8609  -1.6468
#>    1.6709   2.0628  -6.7672  -5.3350 -18.9703 -11.0896   6.4598  -1.1666
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]