Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.1243 -10.4702   7.1645   5.4194 -11.4100   5.5997   1.6044   5.5043
#>    2.3217  -1.3913  -0.7479   5.9554  -7.5108  -7.9729  -2.7127   2.7380
#>    3.4229   4.0261 -10.3680  13.8545   0.9224   3.8379  -7.0816  11.0700
#>    2.1411  -5.7497  -3.3206  -2.8833  11.3721  -8.6199  -8.3983   1.7161
#>    0.7052  -4.6434  -7.5157  16.4955  -6.5977  -7.6725  -2.3853   9.6340
#>    4.2799   5.2793   2.9992   5.3400 -11.1450   2.1490  -3.1709  -2.6385
#>    1.8472  -1.5853  -8.0398   9.2620  -0.7444  -7.3234  -6.1687   0.1404
#>   -6.2514  -5.0189  -3.0672   9.7603   0.0807  -4.6463  -8.5118   5.2171
#>   -3.7857   8.5379   4.0395   0.1687  -8.3244  15.6910   1.7469   7.3035
#>   -1.1072   3.3054  -1.9930   5.3165  -9.8696  -0.8829   2.8713  -2.3113
#>    1.4376  -5.2954  -4.6733   3.8019  -7.2646  -0.6061   7.6615  11.2718
#>    0.9289   3.0411   3.9524   3.4592  -3.1191  15.4325 -11.8247   7.1360
#>   -0.0893   6.1291   2.4322   3.7096 -11.5151  -4.8903   4.4077  11.3847
#>   -3.0244   4.7557   2.1432  -2.0113 -12.1114  -2.1148   4.3232  -0.7462
#>    6.1943   6.9965  -1.0899   1.3629   9.9475   4.9312  -3.5786 -12.9880
#>   -2.4694   1.7180   3.3870   1.3654 -12.8696  -7.6668   1.9845   1.8126
#>   -1.9244  -0.9945   7.2161  -4.1275  11.3279   2.0236  -3.9502  -5.3279
#>    1.8048   5.0983  11.6876  -2.0661 -10.2505  -1.7289  15.6876 -10.7781
#>   -1.5248  -1.3990  -5.7942  -7.8453  -4.6525   5.9652   4.1192  -4.0536
#>   -4.6737  -1.9771   0.3786   0.5530 -11.9058   1.1303   1.8854 -13.0677
#>    5.6751  -8.9527 -10.1781  -0.1958   7.6193  -4.9544  -9.4099   3.0271
#>    5.1242   8.1845   6.6283  -1.3543   5.6334  -8.6170   7.7419 -25.3799
#>    2.2513   5.1417 -12.2568   4.2110 -11.5358  -0.1464 -17.3780   7.8934
#>   -1.0105  -3.5404  -6.7672  -8.3014 -13.2473  -6.6486  -2.0334   9.4555
#>   -2.8885   1.4203  -9.9772  11.1957 -19.9212   6.9356  -7.8244 -12.1150
#>    4.3043  -7.8357   2.1798  -0.5082   6.3957 -19.6113  -6.4479  16.0318
#>    5.2217  -1.1777   0.8803   9.9495   1.4773   3.7163  -7.5767  -0.6555
#>   -1.4870   5.9133  -3.9451   5.0777  -2.1800  17.0793  -1.2000  -2.4694
#>    3.1090   4.8174  -3.3435  -0.3982  -3.4317  -4.2070   2.5509 -35.9639
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]