Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.1004   8.5177  -8.0893   1.3587 -12.2359  -1.8372  15.6696 -15.8540
#>    8.3917   3.8493  -4.3364   5.3823   4.1041   6.2238  -6.8875  -0.4431
#>    7.1628  -1.8854  -1.1027  -0.3195   4.6123  -8.5841   4.4528 -11.3276
#>   -2.1202  -1.7833  -3.3092  21.5510 -13.1156  -3.0433  -0.0969 -15.2109
#>    2.5947  -6.8063  -2.9241  -9.5143  18.1775   6.1978  -8.7269  -9.3126
#>   -1.0672   3.6351  -6.9962   6.7435  22.4119   0.3113  -6.4519 -16.4482
#>   -1.7870  -4.4520   4.3832   3.2974  -4.5234  -1.3802  -2.7061   1.8209
#>    2.2983  -1.5496  -2.4872   5.3894  -1.5597  -0.6883 -18.7257  -3.7756
#>   -1.7458  -0.1931  -2.3366   0.1179  -9.9954   6.7479   5.2558  -8.5102
#>   -1.4998   1.6123   0.6463  -4.9650  11.6812 -17.3662   4.2668  -0.1550
#>   -0.9769  -2.3237  -5.7765  12.1890   5.2770  -1.2337   8.6708   4.4493
#>    4.8175  10.7105   1.2561 -11.7189   4.5494  15.4748   0.5847  -7.9676
#>    1.5948  -7.9359  -7.5986   1.2899   3.1551  -4.0926 -15.0110   2.2821
#>    0.0771   8.3271  -2.2547   2.7364  17.6412  -0.6630  -1.4728 -14.0926
#>    7.2561  -4.9558  11.2806  12.0416  -7.7666   2.1770  13.0430  11.9235
#>    5.0169  -5.5501  -6.3806   3.4163  -3.7380   4.6224   3.2361  10.7717
#>    3.9288   0.4628 -10.1596  19.2812   1.0097   2.4426  18.3999  -0.4373
#>    2.6469   4.6222 -10.1538  -4.8278  -7.0873  -0.0651  -2.7572  -0.6798
#>    1.9754   4.6181   6.7714 -11.6093   6.7388  -4.8082  -3.8332 -15.6945
#>    4.3061   3.1741  -4.4481   2.6776 -13.2740  15.3622   0.4672  -3.7624
#>    2.7062  -0.8555   4.2475  12.7801  18.5149   2.4950   5.9287   4.0074
#>   -0.1462   0.4101  -4.2222 -11.6174  17.0395 -10.5620 -17.5273  15.4465
#>    2.1030  -2.7459  -5.1185  -6.5571 -13.3701  -2.7407  -8.5066   3.6509
#>    0.2416   5.9381 -12.0595   4.5993 -19.2426   2.7462   6.8269 -11.2872
#>   -0.8255  -2.4794  11.9576   0.8917  11.6978   7.9198  22.0694   2.3391
#>   -0.4866  -4.0317  -1.5200   9.1922   4.4927   1.0170  -2.7083   6.6771
#>   -7.7225   1.0558   2.8990  -3.9879  -8.7576  -4.2309  11.4425 -23.9515
#>    0.8467   2.2996  -6.3465  -0.2357  -1.4949  -2.3428  -7.2686  -4.3573
#>   -6.9755  -0.4619   2.3146   2.2332   4.2437  -4.1587   2.8212  -5.8633
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]