Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8 -10.0248  -1.6701  -1.6172  -6.2730  -2.5476  -4.2694  23.7412   8.0710
#>  -10.5266   1.9018  -1.6484   8.7900   7.7552 -13.0623   9.7017  -1.6520
#>    8.8090   4.1768  -6.8287  -1.0954  -9.6291  -1.9214   4.5897  -2.9348
#>   -1.9526   2.1665  -1.7052 -12.1237  -6.5650  -3.2804  -6.7934  10.5121
#>   -4.8140  -0.9850  -0.0455  -2.3906  -0.8238  -2.6390  -3.0345   1.7253
#>   -2.0222   5.1888  13.4619  -5.8952  -5.8090   3.0222   0.6116   1.8667
#>   -5.0988   2.0757  -2.4404   6.1704   2.2201   0.6323  -7.2962   1.9916
#>    0.5736  -6.1405   4.3909  -5.1856  -0.0654  -1.0069  -9.8203   8.3915
#>    5.4433   0.0539  -8.7794  14.9177  17.8458   5.3177 -12.6866  -5.5749
#>   -0.5701   7.9989   0.2179  -0.6648 -17.4453   2.6231  -2.4760   9.4349
#>   -8.6062 -14.0709  -1.6304  -6.0768   2.9132   7.3417  -7.2611  -7.2124
#>   -6.8902   0.9085   9.8981 -10.3669  15.7612  -2.5493   2.0843 -18.8289
#>   -4.3029   0.8901  -9.9673   2.3290   1.8887  11.2116 -10.9725  -6.7019
#>   -8.2151  -5.2342  -5.3715   3.2827 -15.8125 -10.1168  -2.2869   7.9115
#>    6.4697   1.8016  -1.3290   4.3388   2.0402  -1.6296 -22.3361   0.3597
#>    6.4623  11.7934   2.5221  -2.6233  -5.9284  -0.9994   1.2911   3.0677
#>    8.4884   3.0459  -7.2599  -3.8577 -14.4854  -3.4415   6.3984  11.9850
#>    3.0289   1.9943  -7.6608   4.8076  -2.1015   3.5085  14.1733   2.0297
#>   -3.2070   2.4876  -0.0555  -8.4760  29.6121  -5.3978 -17.8988 -14.9825
#>    9.7057  -4.6819  -3.8338   3.7891   8.9171  -9.7222  -2.0545  -1.3406
#>    5.2218  -1.2119   9.4709  -8.9625  -9.7643 -15.0846   1.7496   1.9838
#>    5.8029   1.0968   2.2489  -7.6203  -0.7404   8.1570  -9.5054  18.6904
#>   -3.5705  -4.5765   8.0348   6.9245   0.2487  -9.5756 -14.4487   3.9344
#>    0.3378  -1.5860 -15.8158  -5.8805  14.0452  16.3210   2.8238  -2.8951
#>   -2.7305 -10.9362  -0.3402  13.3856  -6.9581   9.5160   3.0000 -23.9384
#>   -0.8077   2.3541  -3.2448  14.7305 -14.3839 -11.0332  -4.4938  -6.4511
#>   -5.3312   1.4023  -0.4401   6.3092   6.7889  15.3031  -4.2143   1.1105
#>   -0.6328  -4.2187  -9.6374   8.0552  -2.2996   0.4927  -1.4919 -15.3765
#>    1.8121  11.0453  -6.6660  -5.3255  -5.0040  14.4277   3.8322  -0.3987
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]