Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.9485  -0.7628 -15.8507   4.8157  -2.8541 -12.5230  -4.0422   2.2523
#>   -2.1628   0.9605   0.0451  -2.1953  -9.2999  -5.9386  -0.8284  -3.8899
#>  -10.8143  -2.5600 -13.1265  -8.4388 -17.1910  -4.2514   9.0641  -6.9851
#>   -2.3384   1.9308  -2.2118  -0.3003   9.6241  -5.1788   4.0519   7.3577
#>    4.0847  -1.3966  14.2215  10.0688  -8.3672  -7.5017  19.6536 -15.5602
#>    3.9369  -2.0999  -8.6644   0.7132  11.4469  -2.6527  -3.1753   5.3367
#>    4.9482   9.1469   3.2603  -8.1923 -12.0181   5.3812   0.9958 -13.6032
#>   -0.6682   3.1382   9.1944 -17.7338   0.6632  11.4934 -20.8513  -5.5245
#>   -3.0731  -3.8384   5.0583  -3.3044   3.1997  -1.0260   5.0594   9.2362
#>    4.4952   4.8556  -2.8851   0.0432  -6.5834   8.6531  -9.6000 -10.4479
#>    1.5427   5.1030  -2.3837  -3.3741   2.9700   3.2058 -12.7009   6.2040
#>   -2.1906   5.7347  -5.8957   0.5993   2.4425  -8.6883  17.2052  -8.5869
#>   -2.5559  13.7566  -2.5868  -1.7545  -1.2418   6.8256  11.8214  -2.7243
#>   -4.1166  -3.3047   0.4442   1.1646   0.9078   6.6525  -2.3604  -1.7877
#>   -2.8612  -0.1635  10.4176  -0.8140  -2.0310   6.6782   6.6161  -4.4009
#>   11.1301   0.8206   2.6711  12.8437  -4.9581  -1.7540  -4.4237  -9.9881
#>    1.7239   1.0525  -4.9148   0.6454  12.8039  -3.9872 -21.9069  -7.3873
#>    1.5649  -8.6948   3.7671  -8.0403  -8.7472 -18.1067 -10.5955  -3.4937
#>  -10.3636  -3.2048  -2.0866  -2.3540   3.9054  -1.3253  -1.1196   4.5935
#>    0.5994   2.8438   8.4017   2.2704   9.1889   9.1633  -5.1566  -3.0067
#>   -2.7651  -1.6922   8.9162  -3.3509  -1.7718  16.4335   9.7681   9.3531
#>   -9.0177  -9.9073  10.1146  -3.7990 -13.3941  -7.7680   7.3596   1.9741
#>   -0.5078  -1.4586  -3.1509  -3.5191   7.5111  -0.0029  -6.3829   7.3920
#>   -3.3057   0.3652 -14.7303   7.0648 -12.6032  -1.9744   4.4000 -13.4331
#>   -1.3229   0.8460  -1.6116   7.0899 -14.4977  -0.3316   7.3136  12.1867
#>   -2.7830   0.8959   3.8993  -1.1674  -2.4087  10.7421  20.6620  -1.3762
#>    8.3383  -4.4861  12.5915   3.3483   8.1184   2.8574  -5.0631  -3.1423
#>    2.8040  12.3571   2.0005   0.0342 -10.6128  -0.6563  12.2824 -17.5970
#>   -1.4254   1.2853  -1.2264  -0.1512   6.5092   9.5383  12.5145   2.5055
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]