Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -1.4552e+00 -4.2143e+00  2.1152e+00 -4.2894e+00 -3.2255e+00 -5.4242e+00
#>  -5.1987e+00 -7.3660e+00 -1.0970e+00  2.6091e+00  4.8396e+00  3.8577e+00
#>   4.5324e+00 -5.4726e-01 -3.2100e+00 -6.9495e-01 -6.7147e+00 -9.5210e+00
#>  -4.8192e+00 -8.0477e+00 -7.3335e+00 -9.1230e+00 -1.3500e+01  9.4301e+00
#>   3.4818e+00 -4.2251e+00  3.5671e+00  5.8714e+00  8.0004e+00  5.4327e+00
#>  -7.4115e-01 -6.4205e+00 -3.7238e-01 -5.2809e+00  8.2858e+00  5.4556e-01
#>  -3.8529e+00 -8.1918e+00 -1.6601e+01 -5.3319e-01  9.3885e+00  7.2006e+00
#>   2.1451e+00 -7.5367e+00 -8.7403e-02  3.3138e+00 -2.4059e-01 -8.0504e+00
#>   2.0794e-01 -4.5884e+00 -3.4693e+00  1.1728e+01  2.0385e+01 -7.3527e+00
#>  -4.8705e+00 -9.5355e+00 -8.6758e+00 -1.0937e+01  7.1245e+00 -1.2047e+00
#>   2.6386e+00  5.7652e+00  1.1270e+01  9.9627e+00  3.0373e+00  8.2946e+00
#>  -3.7124e+00 -1.1476e+01  2.1661e+00 -3.1895e+00  5.3134e+00  2.1565e+01
#>  -1.8449e-01 -2.6455e+00 -4.7538e+00 -6.1012e-01  1.5641e+01 -1.5077e+01
#>  -3.4652e+00 -6.2520e+00 -6.1627e-01 -7.0744e+00 -1.6070e+01  2.0625e+00
#>  -4.9726e-01 -5.2184e+00 -9.1426e+00 -2.2909e-01  6.9017e+00  9.1073e-01
#>  -1.0535e+00 -1.0318e+01 -2.3290e+00 -1.6432e+01  2.2156e+01 -1.1585e+00
#>   9.6249e-01  2.5970e+00  1.2961e+01  7.4093e+00  9.9733e+00 -1.5773e+01
#>  -4.0732e+00 -6.2599e+00 -1.3901e+01 -1.0028e+01  2.2641e+01 -5.3720e+00
#>  -8.6347e-01 -7.9258e+00  1.8666e+00  1.1570e+00 -1.8313e+00  3.1189e+00
#>  -2.3882e+00 -7.7852e-01  3.2402e+00  6.1536e+00 -9.1137e-01 -7.6873e-01
#>  -1.4401e+00 -1.7350e+01 -1.9638e+01 -9.1335e+00 -8.5422e+00 -1.1881e+01
#>   9.9531e-01  5.9856e+00  3.5548e-02  1.0326e+01  6.4471e+00 -4.1961e+00
#>   3.1303e+00  2.0315e+00  1.1611e+01  2.0331e+01  2.5821e+00 -8.7747e+00
#>  -3.9304e+00 -2.7501e+00 -2.4181e+00  4.8533e+00  1.9892e+01 -1.5773e+01
#>   6.3911e+00  8.8384e+00  5.6559e+00  2.4288e+00  7.8911e-01  8.8443e+00
#>  -1.2138e+00 -7.1214e+00 -6.2626e-01  2.8623e+00  2.8173e+00  2.0929e-01
#>   1.6049e+00  6.2749e+00 -9.9045e+00  1.8000e-01 -4.1648e+00  1.3285e+01
#>   1.8066e-01  7.6321e+00  2.2850e+00  3.6144e+00 -9.0449e+00 -2.8041e+00
#>  -8.8878e-01  4.7304e+00  7.3699e+00  6.3699e+00 -3.3063e+00  1.1045e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]