Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.0415   8.7943   3.8745 -10.5391  -5.8390   3.6781  -7.5819  21.7375
#>    3.6856  -4.3904  -7.6766  -6.9845 -11.5525   2.5029   4.2435  -3.1888
#>   -6.8109  -3.6612  -4.2353   0.6734   0.3209  -5.2118 -19.5834   2.0344
#>   -9.1385   6.1692   2.4100  -0.0478   1.4603  11.6716 -15.8184   2.5028
#>   -0.1837 -12.5697   2.9828  -1.9676   1.7441  -3.4616  -4.6649 -16.4893
#>    2.1810   3.6576  -2.1038  -8.8965 -10.3264  -2.1496  -0.5894   2.6540
#>   -5.8838   1.7706   5.4241   0.8712   8.6008  12.6262   2.3955  -4.8519
#>   -0.7977   7.1801  -6.4989  -4.8595 -10.6645  -0.7001   0.9743   0.5307
#>   -0.5587  -3.0121  -1.2187  -3.1778   5.8786 -15.2366  -4.0862  -8.6006
#>   -1.5629  -0.2820   6.6810   2.6719  -4.5659  17.8754   4.5022  10.4105
#>    6.7949   0.9127   6.6936  -7.5699  18.8517  10.3392  -8.8744  13.7871
#>   -7.9925  -2.9867  14.7184   6.2460  -8.6875   8.1206  -2.1447 -14.7774
#>    3.7274  -8.2572  11.2641  -2.0231   3.8657 -16.7379 -18.2742   2.7789
#>   -1.7089  12.3383  -0.5569   6.3901  14.7685   6.2944  -7.6095   8.7494
#>    4.2454  -1.0265  11.9797   2.2887  -2.9374  10.1700   5.9371  19.6492
#>   -0.5806  -4.4998  -7.5126 -12.7091 -13.4264 -16.6579   2.2770  -9.4672
#>    0.9463   4.3936   6.2890   1.9885   0.2328  17.3264  -8.0812   1.3807
#>    7.4193 -10.7102   8.0901   6.2229  -3.5457   3.3917   2.4798   8.5980
#>   -7.8340  -0.2353   8.3248   1.6277   4.9394  -3.3962  20.4836  -8.5039
#>    0.6452  -1.5587  -4.0473  17.6898 -21.2822 -15.2798  -1.0293   0.7384
#>    2.4539  -1.4317  -1.0129  -7.5147   3.0221   0.6521   2.2996  -7.0464
#>   -0.1613  -6.3563  -2.1013   3.4298   1.8063  17.1260   5.2402 -13.5481
#>   -0.0049   1.8540   5.4560   4.3922   2.1283  -2.0770   7.8529   5.6688
#>   -0.7725  -2.3829   2.3014   7.5101  -4.4156 -11.0188  15.0296   2.9437
#>   -1.1120   7.7587  -1.1212   9.8409  -5.4286  20.4062  14.7019  18.6775
#>    2.7598  -3.3085   3.2457   4.4397   5.9931   2.3054  -3.4283  -9.0594
#>    1.8972   9.9850   0.9874   2.3507  -7.0840  13.1713   6.3874   7.4754
#>    0.8444   2.8706  -5.1810   3.5734  -1.9611  -0.2138  -7.3876   8.2571
#>   -3.3876  -0.1118  -0.0291 -10.8309   1.3690   5.3991 -11.0097  -2.2782
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]