Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.4858  -4.2360   9.3082  -1.1373 -19.4435   7.3263 -17.1126   6.6058
#>    3.6797  -0.7765   1.8305   0.3056 -14.9762  -9.2348   1.2016  -3.3358
#>    2.6628   0.0697  -3.5557   2.8538 -22.1909   1.4631   5.3577  -9.5698
#>    2.4337  -2.0419  -4.9915  13.1163   1.9634  -9.1903  19.4682  -1.7073
#>   -4.8714   0.8888 -17.0736  -5.1852   8.0313  21.1795   2.2158 -14.0245
#>   -4.9998  -0.8863  -5.2884   1.9103   8.5965  10.4267  -8.0871  -4.1410
#>   -3.1575  10.2536  -1.1773  20.5930   0.6931   4.6981   2.2934  -4.6945
#>    0.1727  -2.5125   0.3777  -1.7876   9.7873  -5.7458  -7.6923  10.2566
#>    5.1154  -4.6288  12.6298  -6.7093   6.1069  -7.5330  15.1207 -14.3683
#>    1.0887  -3.6577  -0.2317 -14.8257   4.2417   0.4733  -2.2601  -8.0420
#>   -6.9221   2.1155   4.9676  -9.2152 -16.1539  -4.3192   5.7442  -1.5241
#>    0.2162   2.1300 -13.6148  -0.2495 -10.4854   8.3391  17.9555  14.5855
#>   -1.8985   0.2694 -12.2358 -10.2046  -8.3323   2.3572   3.6754  14.4112
#>   -3.8105   3.2217  -5.4039  11.5345   1.0020  -4.9944  -1.7901   8.5238
#>    0.9686  -5.4498   6.0753  -3.0349  -5.7501   5.7526  -5.6498  -7.4210
#>   -7.6847   2.5080 -12.0217  10.1961  -7.0500  17.0625  -0.7500   4.9518
#>    4.2510  -2.2017 -10.4360   2.5884   3.1713 -16.4042  20.3872  -1.0396
#>   -2.1381   3.4507  -2.1218  -5.5958 -13.0546  -1.9945  14.6117   5.1782
#>    1.2154   2.0655   3.6564 -12.8460   6.3772 -14.7383 -12.3896   4.4720
#>   -3.7955  -2.0496  -9.0641 -22.2761  19.6232  22.1743   0.4093  -4.1561
#>    2.9519  -3.9200   7.0322  -6.5557   1.5154  -2.6468 -18.5896  12.1671
#>   -3.3272  -2.3209  -7.7311  -0.3975 -14.0326   1.9144  -0.6884 -10.4798
#>   -1.2651  -5.3721  -2.8682  -4.6244  -6.0612   2.3170 -15.6192 -19.9520
#>    5.3806  -8.7886   0.3598  -7.5242   2.6148  -5.8194 -11.5067  -9.7664
#>   -3.1480   0.6291  -0.9400   4.6798   3.6490  -1.6310  -4.2586   3.2180
#>   -3.5124   0.4620  -1.4698   5.3575 -11.3282 -12.0686  -4.4157 -19.5364
#>   -5.4028   2.8024   0.7199  -5.8958  17.8168  -0.4833 -15.1433  -5.4429
#>    0.7053  -6.6792  11.7816  -3.4163   2.2655  -3.8315  -8.7374 -15.8347
#>    3.2930   1.2524  -0.4489   1.7684  -0.6311  -0.7480   4.7612  -1.4799
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]