Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.7575  11.6406   0.7748   2.8506   7.9182   9.9797   4.8333   8.6233
#>    9.6563  11.8231   4.7974   0.9155  -5.0012   3.7487   8.8770  -0.9573
#>    8.1661  -6.0378   9.7717   6.1097   0.9898 -18.2489  -7.1629  -1.0904
#>   -3.2718  -7.7887   9.3068  -0.3003   2.4801  -1.9460   5.0419   6.0401
#>   -3.4153   5.8470   3.3214  -1.1041 -10.5429  -1.5823  -6.2684   3.6605
#>    1.5446   2.9287  -6.2426   4.5739 -13.6460 -15.1042   5.7712  -3.0335
#>   -6.0608  -3.3473  -2.2369  -8.7161   7.0053  -1.6098  -4.9138  13.6536
#>   -4.8768   5.1241   5.2625  -3.7950  -9.2873 -14.3577 -18.8498 -11.0303
#>    0.2129  -3.5856  -2.3223   6.2198   7.9294 -21.4221  20.4483  -9.4592
#>   -3.4160   2.5667  -6.4670  -4.9583  -1.8914 -10.9527  -2.4906  11.0660
#>    3.8917   7.4593   1.7513   0.8571   7.5886  -2.2158 -11.3378  -6.7502
#>    2.9127 -10.8970  -9.3847  -7.4218  -4.5711   4.5522 -11.5233  -0.2477
#>    1.5916   3.6392  -3.7848   7.7693  15.8020  -0.9269  -2.3381 -12.6526
#>   -3.1162   6.2515   1.1205   6.1151  11.8579   8.1274  -7.7767  -0.6121
#>    1.2493   4.7563  -5.7392  -4.0480   5.2785  -3.5122  -5.2061  13.5160
#>    5.5446   8.9823  -6.8967   7.4990  -0.0821  -2.8478  -2.0303  -5.2999
#>    0.9719   3.4045  -3.5981  -9.2368   9.6069  -6.0515  -7.6108   3.1263
#>   -2.7081  -6.9003   1.8401  -1.9199   1.3984  12.1370  -0.6292   5.7402
#>   -1.7683  -3.2259  -5.5455  10.7423  -7.6216  -7.3837   2.8786  -8.1418
#>    1.7403   3.5769  -4.6288  -9.6146 -21.4863  -7.6294  -1.1308  -2.2090
#>   -0.6259   4.4646   4.6267  -4.6717 -11.3603  -1.3032   9.0493  -1.9884
#>    0.8936  -0.7423   0.3596   9.6842   7.8818   7.4439   3.1663  -4.1377
#>    3.1685   3.4111   0.0226  15.0703   1.8967   9.0018   7.4820  -1.6716
#>   -2.9540   0.0083   3.1958  -0.8259   6.2021  -6.7833  15.3960  -4.5494
#>   -3.4824  -9.0685  -2.3350  14.2527   8.2689  -8.8238  -3.7005  -3.1142
#>    6.3330  -7.6764   0.0417   4.7530 -12.5889  -4.5596  -1.5820   8.2670
#>   -1.8014  -5.0371  -0.0588   5.2522  13.3803   3.2756 -15.1587  -5.0442
#>    4.1132   1.8711   2.5448   2.2962  -4.2010   2.3189  -4.8790   2.1361
#>    1.4218  -3.2431  -1.3747   1.1605   9.8002  -0.1397  -2.0440   6.3720
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]