Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.2903  -3.2585  -8.7131   2.7097   3.3658  -3.8871  17.8501   8.1823
#>   -1.7640  -1.5125  -2.4232  -8.9416   6.4265   4.7327  22.2834  -5.0334
#>   -8.1593  -6.2840  -7.8174  -5.6593  -0.3963   8.8607 -10.8260  12.3080
#>  -14.1091  -6.0568  -0.5219  -2.3240  -9.1099  19.4433   9.1718  -9.7524
#>    1.7703   6.6577  -3.9612  -3.1356  11.0259   4.9352  22.6227 -14.3838
#>   -3.3320  -3.8379   3.2650 -10.0173 -17.6590 -25.4246 -11.5286 -12.9583
#>   -0.6856   2.7186   3.0068   5.6000  -6.1957  15.2211  -8.7372  13.1689
#>   -1.1998  -0.1353   5.7222   3.2676  -2.3805   5.8261  22.5605   8.4684
#>   -0.6397   3.9147   1.9727   4.7599  -9.7489   3.8604   9.7941  -0.0443
#>    6.0765   7.9103  11.6958   0.5398  -6.5803  -1.2886   3.4473   3.7357
#>   -1.3332  -1.6605  -3.9251  -8.2767   0.0744   5.9556   1.2523   4.7495
#>    4.7567   8.1193   4.7288  -4.4028  -7.9963  -6.7334  -6.4100  -2.6107
#>    1.8197   0.6992  -2.9314  14.2546  -0.0973   6.7739  -3.7569  -4.7670
#>    5.3476  -8.0138  15.1909   0.5698  -9.6239   8.4917  -9.9010   8.9222
#>   -6.2226  -4.4024   5.1523  11.6537   6.2615  10.4667 -22.6304   9.9882
#>    5.0027 -11.2890   4.9259   0.9347  -5.3694  -6.7595 -14.0808  -2.4497
#>   -9.4066   2.0722   2.6110  16.2699   5.5785  11.7314  -6.9298  14.0492
#>   -2.6415 -15.1920  -4.7320   2.9296   3.1810 -13.5642 -11.3846  -0.5284
#>   -5.7370  -1.8448   2.8309   9.3317 -22.5460   7.7479  -0.5978   8.1913
#>   -2.7158  -5.9401   0.7137   8.2585   5.7995   8.3760  -1.8287  -7.2776
#>   -0.0479   2.2347  12.2096   4.6644  -0.4222   4.5633  -6.2992  12.7323
#>   -2.7476  -5.9631   1.5069   1.4336  12.0562  10.0861  -0.8579  11.0198
#>    0.4105   5.8169   5.5087   7.2501   3.7005   1.9712   0.8749   5.5938
#>    0.2810   9.4633   1.5065  13.1308  13.6297  15.8186   2.9983   1.1285
#>   -4.5478   6.5816   4.5335  -6.0548  -0.4096  13.1030   4.9956  -5.1375
#>   -0.2588  19.2630   0.2986  -8.4808  -4.6823  -1.7392   1.5149  -0.9863
#>    3.7449  -2.2185 -10.5293 -10.7578   3.5338   0.3502  16.6058 -20.1266
#>    1.0907   4.7870  -7.0267  -3.5857  -2.3508  -6.8904  -0.8040   2.8317
#>    2.2960  -6.2131  12.5267  -1.5910   4.1589   5.2234 -12.4036 -10.9960
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]