Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.4841   3.5004  10.5210   5.4428  -0.8219 -13.0974   6.0467  -5.2013
#>    2.0012   8.3351   1.5687   4.7059   6.4567  -9.0987   3.4203   5.0022
#>   -0.7185  -4.1713  -0.9508  -3.4756   5.7701  10.2712 -23.3742   1.2611
#>   -2.2638   1.2353  -6.4217  -0.3707   3.0448 -15.0924   4.1180 -11.0530
#>    0.6606  -5.9285   1.9403  -7.0160   4.8543  11.4487  -6.8930  -3.3121
#>   -2.3275  -7.3124   0.2818   0.3783 -22.0665 -13.2509  -7.9123   2.4798
#>    5.3358   4.8816  -9.8593 -15.8005  -0.2019   6.1643  -9.3191  -8.3945
#>    5.9777   6.6521   2.8371  -7.5333  -3.3878  -0.7105   9.1138  -7.3526
#>    7.2966  -4.0198  -5.0288  -5.3350  -1.7716  -4.1083   5.8736  -5.7075
#>   -2.5371   5.2688  -4.2672  -5.9095  -5.6570   1.3483  -0.9170   1.3068
#>    0.4747  -4.4612   1.8269 -17.4338   3.2293   0.1135   6.3720   5.4719
#>    0.7392   3.5479 -13.1740   0.3600 -13.3010  -0.7534   9.6319  10.8737
#>   -4.2453   0.4528   7.3431   8.9862  -3.9629  -1.5103  -3.3944  -2.0940
#>   -3.5340  -2.6106 -11.9683   5.4764   6.0583  -2.3194  -9.0545  -6.3389
#>    2.7872   1.0741   3.3767  14.6417   4.9184  -2.8352  10.0973  -9.3382
#>   -3.3040  11.7341   1.9925 -19.3245   3.2300  -5.5139  -3.1297   1.1646
#>   -0.3574  -8.9983   2.0095   2.2697   4.9435  -6.7982   1.4214  -8.7472
#>   -7.3448 -10.4675   2.6822  12.8603  11.2074 -12.1147   3.7114  -1.2716
#>    1.5433   1.1835  -2.8772   9.8889   8.2344  -9.2904  -0.0900 -11.1080
#>   -5.8577  -7.8400   3.5205  16.6293   7.5018  -5.3875   0.8029  -1.0289
#>   -3.5818  -5.1219   0.9664   6.7814  -3.1376  -5.7549  10.1297  -0.6600
#>    3.1174  11.6831  -4.8478   9.2204   9.5667   6.8635  -0.8882   0.3007
#>   11.5813  11.1761   4.7291 -10.9624  -1.0540   4.8383   2.6276 -24.0027
#>    1.2452  -6.5650   4.5783  -9.2527   5.3258  -8.1819  12.0011   6.0122
#>   10.1284  -2.3675   8.7659  -9.1595  -8.7270  -3.9628  -4.6276   4.9307
#>   -2.4499   6.9131  -6.0750   1.1424 -13.4851   1.2085 -17.5466  -6.6132
#>   -4.4044   1.6305  14.9631  10.7583   2.3665   3.9038  10.2481   9.3041
#>   -4.0066  -7.0785  -3.3439  -1.6596   0.3634   1.5645  -2.1729 -13.8732
#>    2.7173  -8.9778  -7.6280  -5.5713   5.0122  -3.0099  11.5459 -17.6478
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]