Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.2230  -3.5375  -3.9335 -11.0629  -7.9595  16.2382  13.3552  -1.3039
#>   -8.8510  -2.5473   2.9672   6.3494  -3.8543 -15.3999  10.4325   5.1258
#>   -2.8956   6.2111   2.4957  21.4054 -22.4651  12.0295  10.1515  -1.9201
#>    8.1211   7.3585   1.5332  10.3751  -8.8756  -6.8016   0.9434  -1.2627
#>   -1.8879  -1.9528   0.8452   2.7822  10.7702   1.6203  -7.8339  -0.0590
#>    3.4201  -6.5187  19.4865   9.7471   6.2643   7.4280   6.9324   7.7613
#>    0.1587   3.7068  13.3118  -3.5440   9.7859  11.7314  -9.2915  15.5331
#>    3.6928   9.5590   6.4710   4.8211   3.2891  -8.9244  -1.8850   0.9245
#>    0.0572 -11.2942  -4.4038  -9.1188   6.1806   5.1905   0.6700 -10.7828
#>   -0.0497   3.2724  15.4295   0.4801   2.8717  -2.1354  10.1510  -7.5904
#>   -2.7265  -6.9216  -1.7474   6.9599  -4.3499 -13.9184   5.8485  -4.4136
#>    3.2596  -2.1425  -4.5434  -7.0281 -15.2485   9.7497  -2.7777 -16.5565
#>   -4.9955   2.3917   1.7334  -4.9776   9.8920   6.5824 -10.1639  22.2368
#>    1.6212   0.8863  -5.1506  -1.4472   5.0587  -3.1109   7.8403 -13.0471
#>   -2.0949   7.6472  10.6140  10.7275  -6.8800  17.1642 -21.9834   6.3998
#>   -1.8409   3.6922  -4.2026   2.5448   2.4079   1.8826  -7.6310   6.3054
#>   -7.0189   6.9270  13.8161  -2.1915 -16.3139   5.3463 -11.6352  -2.1392
#>    3.3775   2.2058  -0.8108  -1.4371   4.6173 -18.0616  10.0444   0.9296
#>    2.9964   0.5921  -4.2984  -8.4972   3.3723  10.7249 -14.1600   7.7524
#>    0.0224   1.8572  -5.1765  -2.2394  13.5904   4.1269   1.4951  -4.7605
#>   -3.1386   2.8084   3.8249   1.3623  -3.2371  -6.6184   1.1470   7.3279
#>   -1.9950  -2.1935  -1.6249   3.6840   3.5183  -7.0112  -3.1642   4.7711
#>    1.2788   1.4211   1.4027   8.5168   7.6386 -21.3307   3.2901   7.8133
#>    5.8970  -1.9682   2.5643  12.1806  10.5971  -4.0721 -18.3586   5.9608
#>    1.3659  12.0067   6.5275   3.3316  -1.0020  14.1188   4.9065   7.4075
#>   -0.8944   7.0923  -2.6479  -0.9824  -2.0646   1.0490   9.9675  -2.4203
#>    1.4800  -8.2966   2.9350  -4.6831   4.0643  -8.0375   3.7884   5.8063
#>   -3.4492   1.3409  -0.0454  -0.6193  -2.9505   3.5607   2.9095  -4.1121
#>   -3.5788  -4.3708  -8.8893   1.4371 -14.7793   0.9557   9.1658  -4.4425
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]