Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.2793   6.2293  -1.6525 -10.8330 -17.5590  -5.6484   9.7342  -9.1438
#>   -3.0603  -5.1752  -7.2910   1.1812  21.8916   3.8223  -2.4112  -1.3850
#>    5.5894  -4.0846  -0.3481   0.6947  -2.4446  -6.5035 -10.1388   2.2082
#>    7.6731   3.7638   0.8960 -11.0154   9.5525   4.6204  -5.3081   2.0726
#>    0.6165   3.4755  -2.4945 -12.5244   4.7672   9.5724   0.3046  -3.4901
#>    3.2674   6.4019   6.2307  -0.3851   1.2206   1.4918  -1.0346  -6.0064
#>   -8.3605  -6.5406 -12.4350   5.2741  -7.0102   8.0396  -6.1902   0.7950
#>    6.4224   1.2381  14.2638  -1.3839  -7.4436   1.1842 -12.1481  -4.4123
#>   -3.3913  -0.2302 -13.8805  -7.1347  10.0524  -1.7348  11.8216  -5.3208
#>    0.8035   5.8690  -1.4849  -0.6346   7.5789 -16.8721   0.4857  -4.9717
#>    4.0914   7.2767   4.8148  -0.2898   0.8250  16.8801   2.3636 -15.3944
#>    1.7242   0.4136  -9.2509   6.3152  11.5738   4.8438  -7.6242  -8.5763
#>    3.5179   1.2882  -1.6738 -22.9639   7.6655  15.9867  -2.3311   2.3315
#>    0.2042  -3.8551  -4.5278   8.2434  -0.0991 -13.2125   6.1571   5.1198
#>    0.1750  -0.4416   2.5934  10.6075   6.8181  -0.3199  -1.3814  -5.8119
#>   -3.5999  -2.2860  -2.8965  17.7768  11.0610  -6.5069 -14.7307   1.9539
#>   -3.1938  -1.5804  -1.1568  -6.3716 -13.7938   3.0407  -3.5242  -3.9983
#>    7.4068  -1.0181  -7.0163 -10.6978   7.8964   5.5144  -3.0673  -4.4981
#>   -0.5246   3.1198  -0.7358  -2.2691  -0.2025  15.8849  17.7845  -4.1880
#>    1.1695  -5.6171  -3.7358   1.7543  -7.2006 -15.0975  -1.9200  14.5522
#>    1.2773  -7.4007  -8.2637  -9.5832   2.8864  -8.2188 -20.3753  24.3546
#>   -0.2848   5.2341   4.6197   9.6941   7.1488  -0.7698 -13.3029   4.7104
#>   -6.6376  -5.5236   2.9491   5.0382  -3.7480  -4.4983   0.2064  11.4283
#>   -9.6836 -11.9274   0.9391   0.1043   9.5988  20.9062  -2.0729 -15.6035
#>   -3.6289   1.8542   3.3084  -4.1047  -4.2074   8.6694   4.1556 -11.9874
#>    1.5469   0.8043   1.5413   3.5475 -11.6700  -0.5108  -8.6572   1.5059
#>   -5.8415  -4.5726  -5.4168 -16.5953 -24.0414  -2.5719   5.2781  -4.5777
#>    3.9469  -1.4863  -9.3919  15.1615  -0.9142  -7.9605  -5.2970   6.9485
#>    1.5633   0.9843   5.1552  -0.2163  12.6839  -3.1277   1.6956  -9.6559
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]