Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.0147  -1.9438  -4.7079  -0.8856   3.4328   3.0191   2.7505   3.1642
#>    2.8782   0.5379   5.8627   1.8181  18.2465  -0.1806   5.7805  -8.6733
#>   -0.3141   8.5738   5.2206   8.9605   3.3661   0.1142  -4.6271  -3.2011
#>   -4.3667   5.1976   4.7889   1.9236  17.4405  -2.8530  25.7537  -3.1472
#>   -4.2164   0.5177  -2.8595   1.0520   1.7498  -2.2961  -5.3440  -5.3618
#>  -10.2601   4.5796  13.3302   1.7119 -23.8355  -8.7393  -4.0143   2.0860
#>    3.3996  -1.0990   2.2923   3.7485  18.2488  19.1618   3.1855   1.7417
#>    0.2174   5.6308   2.1542   2.9412   3.0825  10.7911  -0.9407  -1.8602
#>    2.4951  -1.9998   9.5879  11.5884   1.7032   6.6285   9.0621   2.8168
#>    5.3787  -8.7548  -7.7293 -18.9596  -9.3266  -1.7351  -2.2987  -8.3316
#>    0.6987  -7.7840  -3.9039   2.7298   0.4296  -0.4167   1.2853   3.7672
#>    0.9802   2.7642 -18.2393  -4.3722  -1.3476  -0.6411  16.0039  10.1636
#>   -2.6451  -9.9554   3.6195  -9.8338  -5.7759  -9.1873 -10.6768  -3.4974
#>   -3.2302   1.0977   4.1884  -5.5629   0.2449   6.5552   6.6803  -1.6566
#>   -7.9755  -3.3715   2.8823   3.1584   7.2190  -1.9874   9.1117   8.0013
#>    2.1432   1.3151  -5.8158  -0.5667   3.0093 -14.6839 -22.6168 -19.7518
#>   -2.4341   4.6860  -3.7387   8.5818 -16.9462   3.9521 -13.8273  -5.0029
#>    6.0110  -6.4141   1.4791  -8.9797 -13.1132  -7.1903  -2.5966   3.8053
#>    2.8947   2.3381  -3.5526   8.6049  -9.9571   4.7621   1.4712   5.3176
#>    0.8794   1.9186   9.0240  21.0350  -9.4025  -7.3952   4.4876   6.3846
#>    7.5096   5.9569  12.3707   3.4578  14.6717   1.2234  14.5223   2.1943
#>    0.8292   2.8368   1.3715   2.7832 -18.1094   3.2705 -15.4278  16.9292
#>    6.9336  -5.3166 -10.4397   6.6180 -14.8187   5.4721  -0.3791  -5.1730
#>    4.5466   4.3575  12.2666  -4.8349   8.1809 -11.3976   0.3443   4.5008
#>   -2.6210   1.7424   9.4994   5.0142 -12.6426  -2.6307   0.0921   7.3684
#>    0.5731 -11.5546   5.3385  -5.8336  18.3400  16.3862   8.9865   3.6868
#>    3.3436   1.5666   0.1454  -4.5760  10.8196 -12.9222  -6.2790   4.8622
#>    2.2821  -6.9183   9.2623  -2.8605  16.2030   8.5205  -3.3031  -7.4456
#>    0.7373   7.5485  14.2412  -6.4565   7.5118  -3.4716   3.4006  15.9234
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]