Conv_transpose1d

Conv_transpose1d

Arguments

input
NA input tensor of shape \((\text{minibatch} , \text{in\_channels} , iW)\)

weight
NA filters of shape \((\text{in\_channels} , \frac{\text{out\_channels}}{\text{groups}} , kW)\)

bias
NA optional bias of shape \((\text{out\_channels})\). Default: None

stride
NA the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
NA dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
NA additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
NA split input into groups, \(\text{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
NA the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=None, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See ~torch.nn.ConvTranspose1d for details and output shape.

.. include:: cudnn_deterministic.rst

Examples