Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.8395   0.2026   5.5777   2.8962  -5.7381  -4.0632   0.0803   9.7191
#>   -5.1558  -3.2881   2.0540   4.9081  14.9221   3.9193 -11.4305  -4.6315
#>   -3.0649   1.2570   8.8975  -7.3332  16.9465  -0.5774  11.9849  -0.1153
#>   -0.1668   1.3250  -0.1845  -2.6787  -8.3342   8.0780   4.7421  -1.3495
#>    2.7666   1.4976   9.7640  30.5945  -6.2307  -4.6205  -7.1484  -1.8644
#>    1.9172  -4.0582  -8.7927  -5.7241   2.1592   0.9067  -0.9200 -15.7673
#>    7.1725   4.8839  -5.5201   1.5390 -17.1074   1.6056  14.0517  -9.1709
#>    5.2057   4.1566 -12.9016   1.2636  -1.5503   6.5540 -17.8733   5.5895
#>   -1.3226  -5.1140  -2.2101   4.3580  -1.5695  -2.5356   8.4348  -2.0854
#>    2.4702   3.2453   0.7105   5.3180   9.0513 -10.8851   8.2783  -1.7959
#>   -2.7698  -5.7974   9.5146 -10.2156  -5.5971  -8.2155  -1.9952   5.5468
#>   -3.5402   2.1325  -5.5861  -5.6085 -11.5766  -3.9009 -15.7963 -15.7883
#>    3.6515   7.6995   1.8722 -11.5126   1.5120  -6.5510   4.7763 -11.4766
#>    1.3569  -4.9959   3.8188  -4.8006   1.4993  -1.5790  -5.3531   3.7002
#>   -3.4467   0.8361   1.6794   3.0571  -9.6437   5.2692 -15.1085   0.0531
#>    1.9809   4.5336   0.1646  -2.1880 -11.8014   7.8173   1.7617  -5.4457
#>   -0.9818   5.8715  -6.1200   4.4982 -13.8686   4.0434  -3.0323   2.6411
#>   -1.4807   0.2841   2.4560   6.1449  -3.1984 -10.9441  -6.1407   4.1697
#>   -1.2766  -4.7028  -0.7370   0.2434  -0.3419   1.8045  -1.5119  -8.5546
#>    1.6237  -0.6154   8.2393  18.2219   5.7162  -5.0355  -5.4936  11.3036
#>    3.0692   3.1714  -4.0752  -1.4873  12.5725  -7.8059   2.0234 -11.4376
#>   -2.8269  -6.7122  -7.6725   3.7361  16.2491   5.4665  -7.9193  14.9156
#>    7.0245   5.6523 -13.4842  -0.7317   0.0679   4.6784  -5.5025   0.6100
#>   -5.8809 -12.3946 -14.0862   1.2021  12.0788  -0.5891  -7.3901   1.3726
#>   -5.4431  -2.0304  -5.9202   7.9654  -5.3566   0.1040  12.2286  -0.1798
#>   -0.4734  -3.2352  11.6765  -6.3361   0.8092  -1.4741  10.4937   7.4706
#>   -1.0099   3.7506   0.7161  10.9291  -5.0119 -10.4199   3.6378   0.9154
#>   -0.7556  -2.4777  -5.9625   3.9553  21.7373   0.6806  -5.7787  -1.5891
#>   -1.5445   2.4085   0.7058   1.1829  -4.6963  -0.0315  -3.2867  -8.9206
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]