Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.0937   3.9665  -8.5046  -9.3543 -13.5912  -1.5244  -3.8809   1.5750
#>    1.1270   7.8427  -7.8365  -9.5857  -0.3614  12.7326   5.1091   6.3383
#>    1.9874   2.9203 -22.6875  -8.3773 -14.2084   4.8237  12.3104  -2.7236
#>    4.4593   7.4051   1.5341   0.0431 -23.6796 -11.4390  -9.3977  -2.3190
#>    1.8579  21.7661   8.9932   0.2620 -25.4637 -11.2094 -15.7243   1.0371
#>    0.2659   2.1961 -14.3830  -2.7730   4.4176   3.7135  15.0902  -2.8575
#>    2.0341  -7.2458 -11.2100   7.4973  -6.0290  13.6758  -5.6886  -3.5611
#>    5.8567  -6.7475   3.6302  12.1229  -1.0907  -5.6210  13.6622 -18.2921
#>    0.1823   0.2645  17.3573  -0.1269  -8.3928   8.5177  27.2125  -4.2820
#>   -3.4255 -11.1083  10.1397  12.7699  -3.6996  -0.3312  19.0924   9.9739
#>   -0.5813  -0.6921  -1.8854  12.7130   1.6105  12.5528 -11.1095  -3.4442
#>   -3.6363  -0.4635  12.7854  15.9478  10.4458   7.4600  -3.9549  -1.6138
#>    3.6121  -5.2588   3.2930   5.2872  17.9179   0.7297   2.4454  -5.9242
#>    0.2472  -7.8899  -0.3887 -13.1053  -1.1280   2.9875   1.9877   0.6705
#>   -2.4842  -6.5097 -30.8635   0.7781   1.9473  -1.3240  10.8059   5.1328
#>   -1.9509   4.4934  -4.7381 -10.5153  -8.9230  -0.8314   3.6557   1.6016
#>   -4.7622  -3.7939   8.4992   4.8885  20.6854  19.8489   0.5418  -6.8281
#>   -9.1308   1.2176  11.4777   6.1180   7.2439  11.2211   1.2610   6.6915
#>    3.1066   0.4199  11.0472   5.6510  -5.4113  12.6455  29.3055  -4.0577
#>    2.7025 -18.2859   5.5506   7.3863  18.7206  -5.2790  10.1478  -2.1081
#>   -7.5659  -5.7304 -18.8169   4.6924  -1.4055   9.4212  15.9401  14.0023
#>    4.9164   0.6627  -0.6417 -16.1741  -5.2277  -0.7576 -12.9299  -1.1460
#>   -3.0715  -7.4463 -11.9208   0.0416  -3.8908  -7.0059   2.6501   2.9627
#>    6.1348  -6.8557  -7.5868  -6.5889  -8.0307  -3.5122   3.3972   4.5117
#>   -1.2780 -10.9759   1.8680 -14.8626  -7.9774  -0.0915   0.0907   3.5299
#>   -1.6782  -0.4119  -5.9445   5.5008   3.1515  -2.7022 -13.7438   0.6499
#>   -1.6692  -8.3948   5.6736  -9.4493  -8.7557   9.0879   3.5815  11.3925
#>   -0.5589  -0.9414  22.7736   1.1182   1.4037 -18.3327   3.0109   4.6150
#>   -0.2754  -3.2443  -5.4853   8.7644  -1.6062   4.3950  -5.3293  17.0834
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]