Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -8.9570  -6.1134 -13.0553   2.7621  11.4334   7.4437  -7.9574  12.3660
#>    0.6341  11.2148   7.4746   7.3639  -2.0349  10.2449  -0.1471  16.4031
#>    1.3661  -6.4757  -2.6417  15.0102   7.2156   1.6510   1.5749   0.7777
#>    4.4918  -3.5973  -2.6321  -8.7608   9.1935  -1.7035   6.5189   0.8479
#>   -0.3108   1.5831  -5.4379  -4.3541 -10.7943   8.0059 -14.8783   4.1150
#>   -2.3363   4.4957  -4.7481   5.8818   2.2421  -7.6915 -19.0545   5.8515
#>   -5.9862  -7.1808  11.6347   6.2365  -4.5663   4.9897 -17.7727 -22.7419
#>    2.3204   5.1226   0.8922   8.5900  -7.8813  -9.7636  15.8263   1.1361
#>   -5.1267  -0.5172   1.7769 -10.3364  -5.0388   4.6282  15.6108 -14.3109
#>    1.2866  -6.7585   1.9979  -2.3170  -5.4726   5.6000   7.6804  -8.8210
#>   -2.4115  -3.2476  -7.6062  -0.1209   0.3856  -7.0752  -2.0560  -7.8097
#>   -2.4651   6.5275  -2.9319  -4.6099   6.2874   6.4598  -0.2106  -9.6598
#>   -3.3220   5.7613  -5.1355   2.8992 -11.2773  -4.9760  -4.1582   8.4605
#>   -3.4288  -6.0721   4.0290  -0.8663   5.2242   2.0220   2.1908  -2.7031
#>   -2.2684  -0.2908  16.9871  -6.1173  -5.8217 -10.1325 -17.3850   5.9921
#>   -3.9098   2.1424  12.2455  -6.7501  -2.5851   4.5668   7.7884   6.0431
#>    3.6390 -11.8515  -1.4075  -3.5014   5.9816  11.0123  -8.5460  16.2429
#>   -6.4964   3.2974  -2.7160  11.9534  -4.2980  -6.1534  -2.7301   3.1868
#>   -6.3038  -0.3879   2.0113  -5.3607  -0.2743   8.8010  -8.9715  -5.0263
#>    5.1385   1.1660   1.7705  -6.7055   8.6583   7.4806  -5.0228   9.2389
#>   -6.2532   7.7173  -2.7726   7.0646  -4.4491   0.4263 -12.3493 -18.6430
#>    4.5521   3.0076   5.3980   9.9938  -7.8165  -7.2916  -0.9319  13.3176
#>    6.5392  -0.8112   2.9466   2.0705 -10.9358  10.4702   3.6838 -12.9458
#>   -2.4992   1.1850   2.4862   7.2752 -18.2755  -2.1982  12.2985  -0.8443
#>    0.0181  -5.9131  -2.3585  -6.1168   7.1584   9.4122  -0.9436  -4.7143
#>    4.1164   5.5464  -2.4283   5.7731 -13.7588   6.0520  11.9648 -14.9415
#>   -5.7595   3.8588  -4.5521  14.3920   1.8356  -6.6264 -11.8711  -3.0150
#>    7.6927  -1.4707   7.1725  -4.4027  -7.8875 -11.6843  -2.5780   0.0966
#>    0.1752 -12.1174   3.7038   0.8459 -18.1449  -0.2490   9.0667  -3.1704
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]