Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.1131   1.4551   8.6406   3.8803  -2.1424   2.3297   2.3167   3.5434
#>   -2.1077   1.2076  -4.5263  -6.2135   3.5619  -3.2937   4.8056   4.4047
#>    1.4984  -6.1874 -15.5448   4.4402  -0.7820   4.2925  -3.5075   0.9268
#>    1.2116  -5.5712 -13.3154   1.7207  -8.5507  13.1984 -10.9588 -12.3061
#>    0.2634   2.7777  11.9529  -1.5566  -3.4032   2.3138  -8.7364   4.3537
#>    2.9872   2.0013   0.7017   4.1708  11.8887   8.1242   7.4870   4.4799
#>    4.5451  -2.4696  -0.4412   1.7773  -2.2422 -10.4325  -4.0438  -1.0188
#>   -1.4929  -1.0159   3.2708  -7.0694 -11.4770   5.9559 -10.8149 -12.1890
#>   -5.2261  -1.7742   1.2513  -1.0214   9.3688  14.4278   2.2016  15.5669
#>   -3.1921   4.8537  -0.6346  -2.1356 -11.7267   7.7460  -2.9782 -17.1903
#>    4.1172  -3.5006   1.3969  -4.4176   6.1164  -5.8450 -14.3022  -5.4348
#>    1.3883  -2.0401  -6.7989   5.2458  -4.7650 -11.0005   3.3211  -7.2376
#>   -0.3324   0.3339  -3.8667  -1.0737  -2.8975   8.7040 -14.0884   5.3071
#>   -3.5569   7.8560   5.3334  -0.8998   9.1839  -8.7419 -10.9003   3.4816
#>    0.8176  -0.0126   5.0117  -7.4588  -3.9013   5.3753 -18.0916  -1.3352
#>    0.7033  -7.8062  -5.9438  11.8280  -9.4213  11.9288 -12.3390  13.8532
#>    0.4406  -5.0118   4.3170  -7.5049  15.4127   1.6481  -8.9765  13.4354
#>   -5.4796   3.9289   6.7274   9.2834  -0.1013   5.6034  16.4668  -1.9706
#>   -6.7625   4.2926   4.4320  -8.3164   6.3261   4.1316 -12.1952   5.3971
#>    0.3765   4.7021   0.1652   1.5323   2.6491  -4.1330   1.1031  10.3699
#>    5.8058  -4.6997  -2.7234   3.6923   1.6487   4.0438   9.9327   9.2321
#>   -1.1980  10.6377   8.4851  -2.8932   9.9120  -3.1228  15.5665  -3.1560
#>   -1.5588  -1.7752   9.6776 -10.9082  -7.1642  -6.9414  -6.1114  -0.4510
#>    0.1148  -3.9685   8.0846  14.2260  -4.2737 -13.4585   7.5583  -3.4022
#>   -0.0668  -6.6840   6.5555  -0.1965  14.9353  -1.5591  -6.3474   2.7989
#>    2.7686  10.4880   0.9262 -18.1660  -9.9664 -11.9572  -6.3579   2.5492
#>   -1.5406  -3.8538 -11.4896  -1.0420  -9.5164 -19.9308  -8.2450  -7.0155
#>    4.8763  -1.7137  -5.5224  -3.5626  -0.0296  -8.9164  -2.6732  -5.6129
#>    0.9524   2.6837   3.5897   2.2599  12.6519   4.7068  13.9057  -4.7308
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]