Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.9062 -10.0503   3.0943   2.9561  -2.9713   6.8740   4.5541  14.0966
#>   -2.8108   0.4769  -3.4291 -14.0952   8.6810   0.9946   3.4265  -5.6420
#>    4.6172  -5.5538   0.4650   1.2643   7.9703  10.1149  -2.5976   9.4013
#>   -1.5541  -6.0154  -0.8962 -11.7035  18.9993  -9.1889  10.9038   1.0875
#>    4.3870 -14.2251  -1.7491   3.5795 -13.9463   2.6953  13.1834   9.5586
#>   -4.8819   3.6677  -0.9937  -9.5202  -7.3307  -5.5452   6.8106  -7.3780
#>   -6.2174   4.1271   7.7775  -6.5296  11.7270  -2.7800   8.7199   9.3795
#>   -4.4608  -2.2434   3.7024   0.1919   6.6301  10.0329   1.8365  -1.7374
#>    3.4180  -5.6031   3.5275 -13.2719  -9.1759  -5.5296 -11.2417   5.0441
#>   -3.3501  -1.4977   9.1710   1.0781   7.7312   4.2083  -2.6686 -12.7415
#>    6.6719  -0.1720 -10.1069   3.5022 -14.8100  -2.6276   1.3781 -12.2660
#>    2.7757  -1.0901  -1.0731  -0.3096  -5.2275  -9.0390  -3.4840   6.0870
#>   -5.2528  -5.1804   9.5893  20.9892   8.5171  -3.1181  -4.9842 -19.9022
#>    5.9352   2.7079  12.7960   7.7902   9.5039  -8.7829 -17.1342  -4.9295
#>    5.0295  -3.6089   1.2081   4.7470   1.7024   4.1960  -1.0318   0.0475
#>   -5.1233  -2.8396   1.9483  -4.0049   0.7757  14.3633  14.7731  -0.9132
#>   -3.8939  12.5730  -3.5111   6.5736  -6.8332   1.2947 -12.8423 -10.1325
#>   -6.1504   4.6934   6.2981   6.3184   8.9332   5.5364  10.9423  -0.2124
#>    1.8831  -4.4456   7.7630   3.2792   6.4086 -15.3984   6.9388   4.6665
#>   -8.4203   3.0189   4.1056 -15.0293  22.0512  -1.0010  12.0915   4.0545
#>   -1.9290   2.7277  -2.9216   3.9282  -2.4384   1.1488  -5.1695   6.3662
#>   -1.0912   2.7387  -1.9726  -3.5576   6.2107 -15.9150   8.7980  -3.6126
#>    2.4631   7.3599  -1.5927   0.4393   5.0868 -11.9053 -14.0105   1.1360
#>   -9.3215   6.0731  -8.4891   1.7633  19.4138   9.6163   1.3953   1.8803
#>    3.6280 -10.4004  -2.0881  -3.3676  -4.1331 -16.8669  19.6703  -1.6981
#>   -6.4718  -1.9172   5.4885  -1.4393  -3.7761   8.0342  14.4052  -3.3554
#>   -7.8889   5.9216  16.8729   2.8837   5.6611  -4.6645 -11.7965  -4.2127
#>   -1.9170   2.4865   1.0013   3.7700  -8.6451  -4.4359  -9.3881   4.1873
#>   -5.2781  -2.8112   3.8372   4.4413   8.3824   8.2047   2.2146   4.3708
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]