Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  1.4107e+00  7.1011e+00 -6.0785e+00  4.4229e+00  6.1791e+00  1.4807e+00
#>  -3.2015e+00  9.8208e-01 -1.7496e+00  9.7491e+00 -1.6729e+00  3.3620e-01
#>   3.5991e-01  2.7645e+00  3.8575e+00 -1.1103e+01  5.1393e+00 -9.0729e-01
#>   1.7220e+00  2.4781e+00 -1.4764e+00 -3.6846e-01 -4.7760e+00  6.2533e+00
#>   2.5675e-01  3.8698e+00 -1.1162e+00  4.0432e+00 -1.3630e+01  5.5579e+00
#>  -5.1082e+00 -5.8769e+00  3.3094e+00  7.6261e+00 -9.6334e+00  9.9094e-01
#>  -1.6897e+00  1.3047e+00 -8.0352e+00  3.3756e-01 -1.8145e+01 -1.3292e+00
#>  -5.2151e+00  8.0930e+00 -1.1950e+00  6.0192e+00  1.3515e+01  3.0045e+00
#>   2.4222e+00 -6.1736e+00  5.6672e+00  7.5827e-01 -3.6116e+00  8.6060e+00
#>  -2.4171e+00 -1.1285e+00 -4.3704e-01  6.8090e+00  5.3027e+00  1.8747e+01
#>   1.5836e-01  1.2610e+00  6.5016e+00 -3.5249e+00  2.0729e+00 -4.2214e+00
#>  -1.6377e+00  3.5076e-01  3.9106e+00  3.5395e+00  1.4687e+00  9.5123e+00
#>   2.3251e+00 -5.3321e+00 -3.3502e+00  9.0538e+00 -4.2148e+00  1.5126e+01
#>  -5.8406e-01  1.0355e+00  9.2603e+00 -7.5616e+00 -4.5209e-01  4.8841e+00
#>   1.7591e-01  3.6472e+00 -2.1844e+00 -3.7976e+00 -4.3520e+00 -2.9993e+00
#>  -3.9055e+00  8.6810e-01  1.2125e+01 -7.1251e+00  1.6659e+00 -9.3575e+00
#>  -2.5879e+00  7.6404e+00 -2.4374e+00 -2.0425e+00 -8.4923e+00  3.1826e+00
#>   7.5203e-01  2.3493e+00  4.5036e+00 -1.1858e+01  3.6729e+00  2.6991e-01
#>   3.2548e-01 -1.2034e+00  6.3850e-01 -7.4872e+00  6.6900e+00 -5.0973e+00
#>   2.8905e+00 -2.3955e+00 -5.4889e+00  8.3922e+00  5.7846e-01 -8.0740e-02
#>   1.6822e+00 -8.1458e+00  4.5539e+00  5.5371e+00  2.4805e-01  4.8271e+00
#>  -3.9061e+00 -2.6150e-01  2.6939e-02  1.4447e+01  5.3938e+00  1.1798e+01
#>  -1.5771e+00 -5.0141e+00 -1.4447e+00 -6.5296e+00 -4.7764e+00  4.7280e+00
#>  -1.0630e+00  4.2575e-01  2.6708e+00  4.9990e+00  1.2562e+01 -5.2506e+00
#>  -1.2578e+00 -2.3279e-01 -4.3026e+00 -8.2080e-01 -2.3285e-01 -4.0394e+00
#>   5.7901e-01 -3.9563e+00 -7.3414e+00  6.0666e+00  1.3679e+01  1.0830e+01
#>  -1.6936e+00 -3.0998e+00 -7.3370e+00  1.6220e-01  9.4413e-01  1.4648e+01
#>  -3.2017e+00 -1.1338e+00  1.0169e+01 -7.8830e+00  1.0512e+01  1.0490e+01
#>  -1.1775e+00  7.2989e+00 -1.1441e+01 -9.8266e-01  1.1500e+01 -7.0175e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]