Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.9855  -6.4033  -2.5579  -2.9443  -4.5396   8.2647  -5.5482  -0.7874
#>    7.2293   5.2260  10.9108 -12.9598   8.4670  -2.2858   2.2683  10.0881
#>    1.5824  -2.3329  -2.5955   4.8895  -0.9680   9.6464   8.8557  -4.2783
#>   -3.2366  -1.5524   7.2095   9.8613   7.6809   5.4205 -10.6277   4.4781
#>    0.1428   6.1773   9.3087  11.8125  -4.7702   1.3518   4.6586  -4.3397
#>    1.2169  -0.9973  -1.1082  14.4152  -6.9746   4.5030   1.5846  -9.7160
#>   -1.0167   5.9755   0.3815  -4.5092  -1.7910   7.6836  -3.0855  -4.3117
#>    0.5340   0.0228   2.8446   9.8993   7.4310  16.4807  -5.3258 -19.7467
#>    4.1982   4.9218  13.0708  -7.7963   7.4665   8.8230  -2.0799   0.9445
#>    0.2026  -3.2802   3.0703   2.3269   8.9984  -0.3894   3.0407  11.8100
#>   -5.9541  -9.6908 -10.1984   7.0480   1.4334  -2.1794   8.7602 -14.3423
#>    1.8602  -9.8116  -1.2555  -3.0084   1.3481  -6.3965  12.0540   5.8119
#>   -1.8598   4.2767  -1.4257   7.6465  -4.4843  17.1073  -0.4580  -7.8384
#>   -2.6259  -5.2802  -2.1944   7.6035  -5.9903 -12.5878   0.9282   2.3227
#>   -0.7972   8.2649   8.3146  -1.6520  -1.1589  -3.7974  -5.2687  -5.3265
#>   -5.8404  -2.9841  -2.1613  13.8627  -0.4638  -2.2748  -7.9191  -0.7265
#>   -1.8523   5.4975  -0.9735   1.0276   6.0720 -12.6394  -3.9863   2.2669
#>   -1.5829   3.4509  -0.4428 -10.2901  12.4929  17.7296   7.8424   0.7209
#>   -2.6240   5.6263  -4.9494  -2.0404   8.0476   9.7192  18.3487   4.9440
#>    8.1917   0.9757  -3.2948   5.8729  11.6322  10.9858  -5.7289 -10.5126
#>    0.0083   4.7741  -2.6980   6.6717  -2.8889  -4.2212  -3.5626 -10.4632
#>   -3.4250   7.1674 -10.5139   0.3868   2.6499  16.6943  -1.9833 -22.3892
#>   -1.0431   0.2755  -1.2447   5.8515   7.2000   1.8788  -5.1178  -3.5532
#>    0.8720  -4.0278   1.4905   7.5207   0.3243   9.0363   9.4301   8.3224
#>   -1.9504  -8.5917  -9.9126 -14.5925  -4.7575  -1.2379   5.8445   4.0962
#>    0.4477   6.6672  -4.4496  -2.9983  -4.2872 -14.4925  -1.1953  -3.8657
#>    5.8734  -7.3795   3.7191   4.2597 -21.7148 -17.9921  18.4544  -8.9525
#>    1.7592  -3.9077  -3.7204  -1.0810  -5.0606  10.6312  13.6437   6.9878
#>    2.5556   2.2103  -2.5193   5.8796  -2.1707   3.9756  12.0717  -1.2722
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]