Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -4.2234e+00 -1.5728e+00 -8.2228e+00 -7.5684e+00 -7.9640e+00 -1.2793e+00
#>   1.2545e+00 -4.0999e+00 -6.3797e+00  1.6772e+01  1.9145e-01 -7.2203e+00
#>  -7.7316e+00 -4.2508e+00 -1.0523e+01  1.1228e+01 -8.5989e+00  3.9309e-01
#>   6.4272e-01 -2.8863e+00  4.7800e-01  6.2075e+00 -9.9185e-01  4.2482e+00
#>   2.0373e+00 -7.7501e+00 -2.6865e+00  2.1377e+01  1.2382e+01 -2.6583e+00
#>  -2.7851e+00 -1.6843e+00 -1.4563e+01  1.5334e+00 -1.4161e+01  7.2858e+00
#>  -1.5629e+00  2.3664e+00 -8.8420e-01 -7.6988e-01  7.7567e+00 -3.1646e+00
#>  -6.4321e+00  2.1038e+01 -1.0008e+01  6.5032e+00 -3.8968e+00  3.3656e+01
#>   7.5822e+00  7.2633e+00 -1.2110e+01 -1.5321e+01  1.4125e+01  2.5312e+00
#>   1.6173e+00  6.0008e+00  9.1424e+00  5.2582e+00 -9.8981e-01  8.5292e-01
#>  -3.1348e+00  7.7391e+00  1.7419e+00  1.9976e+00  4.4158e+00  5.2534e+00
#>   3.7076e+00  9.0654e+00 -3.4474e+00 -8.9252e+00  1.2565e+01  1.2945e+01
#>   1.9891e+00  2.0812e+00 -1.1380e+00 -8.6408e-01 -6.1325e+00 -2.5447e-01
#>   1.4057e+00 -1.6655e+00  1.7528e+01 -2.1477e+00 -7.4752e+00  9.1115e+00
#>   1.8802e+00 -5.1250e+00  6.2253e+00  1.7814e+01 -6.5107e+00 -1.0081e+01
#>  -6.2037e+00  4.3218e+00  5.6447e+00 -1.0116e+01 -5.0242e+00  8.5651e+00
#>   2.3823e+00 -3.2953e+00  9.6943e+00  4.9657e+00 -1.1000e+01 -6.9641e+00
#>  -2.1691e+00 -4.7182e+00 -3.3008e+00  2.4498e+00  9.4541e+00  6.0243e+00
#>  -5.3326e+00 -9.4787e+00 -6.3699e+00 -1.9890e+00 -9.0547e+00 -1.0267e+01
#>  -8.4583e-02 -1.1812e+00  1.8066e+00 -1.9460e+00 -5.7787e-01 -1.2306e+01
#>   3.1435e+00  5.5832e+00 -3.6380e+00 -2.1087e+01 -1.5307e+00 -1.6790e+00
#>  -2.9184e+00 -5.7900e+00  1.2403e+00 -1.3150e+01 -8.5477e+00 -1.0715e+01
#>   2.0355e+00 -5.1539e+00 -4.1286e+00 -6.4777e-01 -7.6465e+00 -2.5658e-01
#>   1.7781e+00 -3.9021e+00  7.4880e+00 -5.1928e+00  8.1165e-01 -1.2633e+00
#>  -1.0430e+00 -9.6934e-02 -3.5253e-01  5.7990e+00  3.8174e+00 -2.1668e+01
#>  -4.5146e+00 -8.1492e+00 -1.9255e+00  5.1341e+00 -8.7045e+00  1.8021e+01
#>   2.9255e+00 -4.5194e+00 -6.4421e+00  9.9677e+00  5.9384e+00 -1.6460e+01
#>  -1.1239e+00  9.5482e-01  7.6341e+00  2.9196e+00 -1.0281e+01 -3.5739e+00
#>   5.1399e-01  4.9088e+00  1.0340e+01  8.1244e+00  6.2961e+00  6.5787e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]