Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.5308  -9.1517  -3.3052  11.4686  -1.4764 -10.2058   6.5782  -6.8602
#>    0.8493  -3.1681   5.0416   4.6742   0.6241   0.8981 -11.1610  11.1778
#>    0.2756   1.7848  -1.1155  10.4818   5.5371  -8.8168  13.1931  11.0028
#>    4.8852   5.0199  -4.5916   7.8786 -11.0505 -11.4617  10.3880  -6.2733
#>    0.4969  -2.2930  -3.4517  -3.0556 -10.4375  -8.3299  -8.8560   4.3743
#>   -1.4695   2.6999  -2.4049  -6.5846  -7.8194   5.7722  -2.0964 -11.3351
#>   -1.4999   1.4278   5.3692   1.5067  -9.1947  -2.2194  13.4199 -12.0352
#>    1.7531  -4.1730  14.4967   4.1622  -6.7433   6.6065   0.6555   2.5805
#>   -2.7261   3.4444  -2.4879  -6.7810 -15.2541  -0.2624  14.0853   1.8436
#>    0.5104  -4.9994   0.7430  -3.7944   1.2807   2.6096  -4.7909  -2.8664
#>   -4.1121  -1.4268  -1.3745   2.3200   8.5611 -13.5998   5.8827   8.7488
#>   -2.1403  -1.0207  10.3980   4.5334  -1.7831   0.4891 -10.3681  -3.6904
#>   -1.5456  -3.0435   8.6750   4.9320 -16.8356  -7.9034   0.7310   1.9098
#>   -0.1675   6.6727  -5.6989   8.4693  10.2330  -2.5759  -4.3546  -8.2649
#>   -4.4075  -1.7331  -1.1797   0.6326   5.9420   8.7248   6.8817   1.5972
#>   -1.8887   6.7787  -2.3495 -11.6149  -2.0471 -11.5077   0.4417  -5.2356
#>    0.8759  -0.7631   0.2254  -0.4759  -3.8042   2.0153  12.4372  22.5638
#>   -0.3961   3.0783   5.6569  12.2471   5.7679  -8.3295 -13.3767  -0.3067
#>    4.2776   1.3368  -2.3875  14.5298   2.8713 -12.8317   0.7375   3.6248
#>   -1.7617   1.5704   5.8890  -3.3732  14.4163  23.0001  -5.6879  -8.8600
#>    0.2911   5.3183   4.6142 -10.9513  -3.0369   1.6924  12.0812  -6.1676
#>    2.0856  -0.0064  -6.3659  -3.1211  -1.1874  -1.2049 -12.5063  16.9135
#>    1.3393   3.7964   1.2225 -11.8759  -1.5770  -8.4210 -19.6432   6.6711
#>   -0.3027   2.0419  -4.6020 -12.1665  -6.7930  15.3236 -16.9440   0.2232
#>    3.9521   3.9915   1.3220   5.6546  -2.8313  -5.9288   3.9790  -0.6365
#>   -1.9517  -0.0979   3.4909   7.6196  -6.7886   8.1430   4.0715   6.6499
#>    2.5322  -5.7885  -2.2506  14.0052  -2.0524  -9.3463  13.0081  14.0132
#>   -0.3049  -7.8703  -8.7533  -6.1203  -4.2647   4.2606  -1.4151  10.6806
#>   -1.3267  -6.3875   9.2926  10.1706 -12.1312  -7.2660   5.1524 -12.4887
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]