Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.5969   3.5322   8.0531 -11.9996  10.0405 -11.6344  -2.7728  -1.1324
#>    2.4396  -2.0765  -2.1105   4.4214 -10.5549 -21.0383  -7.7659  -2.6131
#>   -0.2734   2.5337  -1.6943  -3.1594  12.7020  -3.7636  -4.4517  13.8654
#>    2.8208   7.2621   0.0820   0.4580  13.0026   9.4525  10.4696  17.5486
#>   -1.5922  -7.6480  10.3466  -4.5613   9.4108  -6.2922 -10.3917  13.0905
#>   -0.1043   3.7127  -1.8798   3.9582  -2.7890  14.5613   7.3096  11.8244
#>   -1.5351  -5.4537  -8.6745  10.6023  -7.8726   7.8522  -3.6961  -3.3148
#>    4.1172   1.4890  -2.3587   9.2681   1.0793  -6.6130   1.4681 -10.0134
#>   -0.4827  -1.9862   6.9969 -10.1847   9.0342  -5.8963   3.9587   8.7632
#>    1.6618   5.4038   2.5010   1.4641  11.0842   3.0689  -1.6846  10.3806
#>    0.4677   0.7941  -0.9389  -8.7367   0.6778  -8.6599  14.2328   7.2611
#>    4.7683  -1.3331  -3.4942  -9.1997   7.9449  -8.6460  18.8458   5.6566
#>    0.4371  -5.7189  -1.6091  -4.8222   0.9200 -18.8162  14.3218  -6.3049
#>    0.5939   1.9656  -2.0257   3.2130  -7.7373  19.6332  -5.3270  11.7962
#>    2.3477   2.3846  -8.7438  11.9200   5.7754  14.5139  -5.2618   8.2266
#>   -0.8121   3.6786   0.8339   1.9575  -1.2312  -6.2005   2.2361  14.9874
#>    0.3241  -0.7232  -9.3551  12.1149  -4.7735  -1.4346   4.7200  -1.1615
#>    0.6072  -4.3592   5.8315 -14.5431  -8.5248   0.0417  -3.7952  12.0925
#>   -0.8369  -2.4767   8.4361 -12.0603  -8.0642  -0.5459  18.7819   3.7851
#>    3.6686   6.6928  -2.6431  12.2411   3.6905  -4.2855  -0.5963   3.7135
#>    1.9386  -1.9632  -4.7411   0.6300   4.9301   3.4400   1.6961   1.9809
#>   -2.6522  -2.7089  -7.3947   7.4857 -16.7151   4.2230   4.7972  -9.2248
#>    1.9280   2.8891  -1.2233  -6.8046  11.1687  -4.5713  -1.7522  17.4684
#>   -3.7854  -5.1800 -15.4053  13.1591   4.8101  10.1920  -8.1750   6.6302
#>   -1.3345   4.9910  -3.1429  -0.0693   5.9020  -3.6328   2.0303   4.4108
#>    1.9235   8.2120  -4.0581  -6.7661  -1.9045   3.8638  -8.9156  12.0840
#>    5.1532  -6.6556  12.3519  -7.0848   4.0539  -4.2347  -7.0605   2.2869
#>   -2.0068   3.4494  -1.5230   7.1646   3.7801  -8.6811  -0.7286 -13.0712
#>    2.1953  -5.0532  -9.1452   3.9083 -20.0717  -1.7974  16.1633  -5.5025
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]