Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.3841  -0.0448   5.5705  -5.1998  -6.3134   8.3620  -0.2314   8.2885
#>   -1.2110   0.7353  -2.4255   5.2604 -17.8808  19.2919  13.2189   7.7762
#>   -0.0349  -1.1951  -2.7630   5.8709   3.6123 -24.2575  -2.7692  18.0325
#>    3.1415   3.0995  -3.6117  -4.9396   1.6030  -3.7903  -0.9532  -6.4249
#>    1.4413   0.5309  -8.9036  -9.0074   1.5255   9.0093   8.2041  10.4767
#>   -2.2136  -1.0668  -7.7768  -3.1920   1.0500  -0.3376  -4.0028  -9.5346
#>    3.6443  -0.4690   8.3495   9.1378  -5.6595  -2.2972   3.3979   8.8257
#>   -3.6786  -6.8338  -0.7361   8.2572   0.7249   6.6049   8.5584  -8.5639
#>   -3.9672   0.7801  -1.2764  -3.4702   7.3864   8.4166  -8.3024   9.5257
#>    2.4612   5.3214  -3.3719 -13.5653   0.5513  -1.1669   3.8018  20.5903
#>    4.2120  -1.9884   2.2087   6.8767  11.8519   1.3908   5.5438  -1.7975
#>   -0.8602   5.7043   1.8066  11.5912  18.7029  -0.7334  -1.9824  22.4315
#>   -1.4002  -0.1817   1.7102  15.5511   0.4569   2.3580   4.5558  13.4206
#>    3.9420   2.9324  -5.1325  -0.6181   3.8707  -7.0796   4.4359  -8.9075
#>    2.6797 -10.5678 -11.1506   2.5640  -8.9576  13.1269   1.1273  -4.9123
#>   -0.2195 -15.8104   2.8517   0.5812  -1.6627  11.1359   1.9103   4.5622
#>   -1.4159   6.3105  -3.8380  -2.4391  11.4392  20.0076   4.2833   9.1906
#>   10.2497  -2.6402   1.3882 -17.8360 -14.4466  -6.5217  -1.4995  -4.3423
#>    3.5696  -0.1364   1.9219   1.7689  -8.5337  -8.8581  -2.0852 -14.0187
#>   -1.9554   5.0370   7.2320  -1.3985  -4.0994   5.4704   6.2419  -9.8424
#>    1.0183  -4.0242   0.8755  -2.2070   2.9774  16.4884  11.0393  -1.5538
#>    0.4615   6.8320   5.1753  -2.3742  -9.1486  -3.4517  -3.5571  11.2353
#>   -4.6790  -3.6994   0.9477  -3.5176 -23.5291  16.4518  13.2704   4.9587
#>   -1.6807   1.9807   3.9688  -7.4171   1.4989   5.4405  10.8950 -10.4911
#>   -0.6690   5.1696   3.3773   6.7215  -6.7893  -5.6739  13.2417   2.2112
#>   -7.2364  -4.0017   8.3846  15.6611  -4.5354  -3.8903  -0.3076 -12.6024
#>   -0.3606   6.7789   2.3135   2.5900  -7.2945   8.6917   2.3024  -0.7469
#>   -2.1261   1.0734  -4.0103   4.1269  -9.8045 -15.6200   5.7372  -5.4673
#>    2.9005  -1.9945   9.0730   6.9206  -9.0096  -9.4763   1.0805   1.3081
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]