Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.7044   2.2838   0.8719  -4.9943  -3.0309 -10.5385 -15.4888  -6.2307
#>    5.6411  -0.6640  -8.0335   1.4453  -4.7556 -11.9635   6.6842 -18.1807
#>    1.2788  -7.9797  24.3520   0.0856  -3.0805   6.4432  13.3611  -1.3935
#>   -8.8898  11.5876   4.0488 -10.2510   3.4120   4.9752  15.0701  -2.0886
#>   -4.8496  -4.6869   5.8544  14.9842  -7.0646 -18.8442   1.1228 -14.2677
#>    0.8443  -0.2239  -5.8725  -9.6944  14.1177 -13.2877 -17.8115  -0.5069
#>    6.0049  -3.7748  -8.9038  -6.1707 -10.4317  -0.0057   5.0120   3.7683
#>    0.8843  -6.6328  -1.2692   2.5236 -15.1246 -13.0856  -2.1879  -5.7933
#>    2.7400  -2.1660  -1.4084   3.9193  13.9790  -2.8054   6.0882   8.7637
#>   -1.3442   0.3503  10.8861   2.2459 -12.5103   5.1212   7.7507  17.4964
#>    9.2459  -9.0150  -7.4886   4.0963 -10.3592   5.8737 -14.8454  26.2305
#>   -1.3340   4.6703 -11.1023  11.1295  -5.9405   0.5809   9.9634  -7.5800
#>   -1.4815  -3.7378  -4.4066   5.1407  -2.8944  -9.2534  18.0047 -12.6459
#>    3.4346  -7.8183  -2.2240  17.7064 -20.0460   5.9194  10.3086 -20.4747
#>   -1.2826   6.3659   5.3966   2.4934   0.6097  -3.7129   2.7616  -7.4679
#>   -2.3833   0.0785   5.2370  10.8958  -5.8815  10.1693  15.4341  -1.7661
#>    0.4507  -7.3385  -3.7995  -1.8060  18.7876   4.6243  -8.8880  -5.5053
#>    0.1910  -3.5285   4.0698  -6.9375   2.7250  -7.0838   6.3694  16.6720
#>    3.0441   1.7411  -8.2720   5.4584   4.8125   1.7323 -18.5082   0.6383
#>   -0.9993   5.5094   6.2544  -8.0314   4.1124   4.2663 -16.4720   1.0321
#>    2.3918   3.7332 -14.5524  -6.2808   5.0322  -4.2450  -8.0399  17.3237
#>   -0.6261  -2.7352  -4.5079   1.7827  -1.0174   2.0160 -11.8265   5.4494
#>   -2.3248  -2.3642   8.8513  -2.2644 -15.8783  11.3377   2.1899   3.5040
#>    4.6717  -2.6667 -12.9064   8.2859  -8.6157  -9.5847  -3.9385   4.7553
#>   -3.7500   8.3492   2.3802  -0.6346 -10.4028   6.2056  15.4284  -2.3400
#>   -2.0800   9.6766  -7.1070   5.3973   8.4054   8.3064  17.5958 -13.3856
#>    1.5201   2.9557  -8.7762  -5.8161  -1.0115 -10.8836  22.4341   5.2744
#>    4.1344  -9.2095   5.4333   2.7247   0.4916  18.6782 -13.0295  -8.3363
#>    1.2592   0.0476  -2.0229   2.7590  -1.2589   4.4793  18.6698   0.7914
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]