Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.2433   5.9721 -11.9456 -11.7034  -5.3008 -10.4728  -8.6155  -9.7777
#>    4.6161  -6.2064   4.2416   6.8041   3.2396  -8.5636 -16.6128   0.8363
#>   -0.3982  -1.3527   6.3225  -2.9014 -13.5408   7.4865  -5.6884  12.5674
#>    6.5335 -12.5097  -5.3740  -3.2929 -13.9031 -12.1107  -8.3062  -9.1435
#>    4.1400  -5.5622  -9.1961   8.1724   8.9439   6.9519   0.0064   3.0993
#>    0.2107  -6.9954   3.4598 -20.9660  -0.0938  11.7507  -8.0990   2.3025
#>    2.0819   0.8126   5.8724  -3.3098  -6.2263  -6.1956  -0.3176  -0.6373
#>    7.7456  -1.7981   4.3843   5.4571  -6.2919  -3.0947   2.1523 -16.7025
#>    2.1602 -10.7315   6.6644   5.8419  -2.1157  -9.5476   2.7621  13.5262
#>    8.7720   1.2502  -7.6587   1.8827  -7.4798   8.6606   5.5149   5.8514
#>   -4.0086  -1.1825   0.9193   4.7254   7.6566 -21.6735 -12.0412  23.3797
#>   -2.3766   8.6325   3.3626   6.7517  -4.6420  14.3854   8.7175  11.6130
#>   -1.3549  -1.1429  -4.2378  -3.5319   3.3315 -10.1500  17.4603  -4.8883
#>    1.6968  -2.3856   3.1700   8.4552   1.0351   1.4486   5.4556  -0.1032
#>   -0.3646  -3.0635 -12.5680  -2.7210   1.2569   1.8099  -6.6666   6.6135
#>    2.6757   0.8228  -4.4640  -1.8212   6.7048  -8.1502  -2.3362   5.3552
#>   -2.0685   0.1163  -3.4639   7.8249   0.1054  -3.3827   4.7583   1.0330
#>   -4.7259  -2.2829  13.0945 -15.8959  -1.2107   8.5573  -8.5585  -7.3478
#>    7.1382  -2.4465  -8.1086  19.1324   2.5603  -7.4917   9.1708   0.0688
#>   -4.5926   4.1799  -6.0429  -9.1290   3.2384   5.9472  -2.6025   6.2484
#>    4.6565   0.8243  -5.8196  -1.3659   2.1370  -0.0900 -10.5894 -12.1243
#>   -0.4908   3.9510   0.8865 -22.3388   7.5721  -1.1706  -0.8733   2.7315
#>    3.9358   0.0856  -2.3148   1.6720 -19.1296   0.5464  -5.5018 -10.8124
#>   -2.5008  -5.1314  -3.2044  -0.5123 -21.0334   9.0598   0.8920  -2.6036
#>    2.8503   1.1240 -13.5121  10.3680  -5.2251   0.1100  25.0284 -10.0111
#>    0.0629   4.9443   0.5724  10.2922   1.1394  -7.9367   2.2937   8.2372
#>    0.3891  -3.2272   1.5470   6.0167 -11.1825  -2.0079  -4.4398   7.9401
#>    1.7729  -8.0440   0.2319  -5.2243  -9.9388  -8.4954  -9.2304  16.7399
#>   -2.4365  -2.6259   7.7370  -8.6556   3.7312   7.9553  10.1952   8.9898
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]