Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.5123  -4.3535  -9.6027   2.5283  -8.0156   2.3896  -7.0898  -0.5258
#>   -4.1399   3.4528  15.0372   5.3989  -5.5404 -11.3454  25.4844   3.0759
#>    4.2605  -3.3567 -16.1230  -6.1786 -13.1767  -1.3322 -15.1719  -0.1953
#>   -2.4070  -2.6605   8.6881   9.0337  -1.1493 -12.9331   2.9246  -4.1127
#>   -2.1911   1.9032   5.3585   8.5500  -6.8149 -12.6479   2.8251  11.6187
#>   -1.0975  -7.6712   0.1047  14.9349   0.2754  11.2766  14.7944   1.6902
#>   -3.2163   5.7240  -2.1251   2.0042   5.7750  -5.2888   3.0339   5.2448
#>    3.7000   2.4196   2.5460   4.7464  -4.9783  -0.9275  -2.5939 -20.7107
#>    4.1855  -2.4270   5.9651  10.8081   1.5269  15.4313  11.9672  14.2521
#>    0.7702   1.9313  -5.8283   6.3045  -0.1164   7.8499   0.4560  -0.5168
#>   -0.1777   2.0970  -6.4657   4.2009   9.2621  19.5078  -1.1904  -8.1796
#>   -0.9277   4.5767   9.2664  -0.0590  16.4515  -2.3408 -12.0048  -0.7665
#>   -6.1279  -3.4348   0.4431  11.2396  15.1317  -3.0100  -6.5708  -0.7131
#>    3.2546  -6.1429   3.5463  -6.7240  -6.9932  -6.7448   2.1185  -6.3858
#>    0.4618  -7.6010  -4.2727  -3.6216  -3.7794   0.6747  21.9495   0.4575
#>   -3.7022   3.1680  -1.5775  14.9306   9.4085  15.1588 -16.9832   5.9796
#>   -3.4880  -1.6448  -7.4118  -4.7873  -3.4622  12.3051   4.8829  13.6232
#>   -3.2606   2.3524   5.6877  12.8741  -6.3452  -7.8442   7.3168 -17.9235
#>   -1.2063  -4.1823   7.5801  -1.2679  -7.8554 -19.5009 -10.0603 -14.6349
#>   -3.0562   2.3681  13.0152   7.1421   1.6658  -2.3946 -19.4387  -4.6013
#>   -4.0760  -0.1605 -10.2138 -10.8001   7.3673  -5.4999 -10.9514   2.0180
#>   -3.6691  -3.3613   4.7136   5.9649   5.4762  14.4597  14.6872  -5.1899
#>   -3.9289 -12.6945   4.3031  -4.5161  -4.0359  10.9215  19.3480  -0.5475
#>    3.6641   7.1547  -2.8602  -0.2547 -17.8790   5.1115   9.3027  -0.5077
#>    0.0185   3.7227  -9.1522 -13.1306   4.0127  -4.5534 -25.7934 -10.9743
#>    3.7974   1.8786   5.5545   0.6954   1.8775  11.8238  -5.3862 -13.7451
#>   -0.1448  -2.6096  -2.4396  -2.4349   0.2484   4.0450  -1.2829  12.0201
#>    0.2924   3.3170   7.4425  -5.6958   8.7290  -5.3547  -6.3098  -5.1176
#>   -0.2800   9.9940   7.6290  -5.3104   4.8042 -22.7239   0.1260   6.4517
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]