Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.7597  -4.0144   0.2226   0.8356   5.3517  -3.4764  -1.5196  -1.8715
#>   -3.0200  -2.5364  10.5252   7.8086 -17.2549   4.2966  10.9719  16.2926
#>    0.8973  -3.7613  -4.8147 -10.8707   0.2826  10.4583   6.3907  -9.4575
#>    1.1966  -3.2505   1.0129 -13.3138  -5.4315 -25.5585  -3.3232  12.1755
#>   -1.5703  -2.8725  -8.7699   6.4805  -8.1806   3.2045 -16.5467 -10.8032
#>    0.7152  -5.4644   8.2747   1.2873   1.2804  -2.2802  -3.0125  -2.3223
#>    1.1875   3.4700  -0.4087  10.1332  -0.3545  -2.0672 -10.0163  -7.2427
#>    0.2042   3.1199   5.8247  -0.3167  -3.9296  -3.9804 -22.0864  -6.1959
#>   -1.8379  -8.1960  10.3926 -10.7029  -4.2372 -19.2054   3.0645   4.6665
#>   -3.2250  -1.6122  -3.1149  -2.1835  15.8551  16.8905   3.3565   1.5601
#>    5.5420  -3.9971  -2.6429   7.2702  -4.5557  -1.3891  -0.6331  15.2484
#>  -11.0776  -2.7022   6.0140   3.9832  -6.8698  -2.8484  -2.7219 -19.4847
#>   -1.7953   3.0518  -2.5538   4.1888   9.5114 -10.7017  12.5935   3.2500
#>    0.6007   5.5024   8.6160  -7.4967  -8.1147   6.7162  10.5079  -0.1456
#>    0.7920   0.3049  -3.0493 -10.2567  -6.9035  -0.7225   5.6525  13.3262
#>    4.0937  -7.5414  -0.7064  -1.6759   1.7514  17.8475   4.4491  17.6637
#>    0.2212   6.3669 -11.1447   2.3060 -11.4868   6.3777   0.4177   9.7112
#>    2.0805  -3.5651  -1.7357 -14.6261   5.5209  -3.2563 -19.1810 -13.0381
#>    0.2586   1.1343   8.0894  -2.2157   1.6198  -3.4894  -1.0945   7.7120
#>    1.5665   0.3352   5.6139   2.1449  11.2757   2.5206  18.5678  -1.2535
#>    2.9329  -4.2189   1.0990   7.3535  17.5026  -2.2341  -0.8747  -0.9952
#>   -1.7654   4.0426   5.3536 -17.2190   0.6358  -7.5284 -10.0941  -3.7192
#>    0.7839   2.8758   2.5781 -11.0748 -13.6933  -5.8628   0.2102   2.1604
#>    2.5251   2.5461   4.8395   2.6030  -6.8728  10.4271  -0.9276  13.7532
#>    0.2270   6.4126   1.8875   1.9832  -9.4275   7.4832   3.1570  -5.7876
#>   -0.4971   1.8872 -12.2848   3.4786  -7.4645  -5.0366  -8.5021   9.2897
#>   -1.0040  10.5097  -3.6528  14.7195  -4.1361  -4.5136  -3.1288   7.2283
#>   -0.6544   2.5491  -4.5253   1.3825   6.5308   5.5055   3.5957  11.6813
#>   -3.6696  -0.1512   3.8793  -8.9948  -2.6400   0.2568  -3.1506   5.5688
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]