Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.0260   3.2411   5.3447   3.8461  -9.6711  -7.5917   4.5669   2.0350
#>   -2.9675  -0.3117  -3.1940   2.6687  10.3777   0.4960  -9.3625   9.3988
#>    6.5797  11.6112  13.0892  -4.3735  -5.4744  -4.6512   8.6741   5.1019
#>    3.4515  -0.3992  -4.2876  11.6532   4.3635  -5.8865  -1.2690  -9.1595
#>    4.2135   2.3266  -2.3978   3.4895  22.8039   4.5272 -22.0184 -10.3754
#>    1.5467   7.5227  -0.7576  -2.4307  11.8796  13.2634  -2.4072 -17.8903
#>    0.4875  14.7774   3.8511 -10.3541  10.5659   4.2040   6.5326 -21.5535
#>    1.0759  -8.3458  -7.8408   2.5260  -4.3769   3.6252  -3.1224   2.9163
#>   -3.9770  -1.7790  -8.1093   6.3143   8.1525 -13.8595   6.8791   9.3331
#>    5.4691   4.1126   8.2214 -14.3204  -4.5854  12.7894   4.9524 -10.6689
#>   -0.1090   1.0638   0.9770   1.6430  -0.7300   5.5443  -0.6784   4.9636
#>   -0.8858  -8.7584   2.7046  -5.2089 -17.9452  -5.1068   7.7962   0.3150
#>    2.7260  -1.9336   1.8013  -2.5317   3.1734   8.6129  12.3569  -5.2186
#>   -0.3717  -9.7155  -1.2131  -2.7792 -10.6060  -5.1788  -2.6725  -8.4219
#>   -1.6732   4.0500   4.3622  -5.6394   6.3157   2.3980   5.4997   2.8215
#>   -0.7976  -2.7590  16.4823  -9.1183  -9.8921   4.7139  -6.3824   1.5836
#>    9.0757   2.4718  -6.6090   0.5589   7.1101   6.9226   0.3746   0.5652
#>    2.0697  -7.3925  -6.4570  15.4641 -13.2483 -16.3003   1.4635   1.9463
#>   -4.3414  -0.2040  -6.9927  13.6233  -3.6942  -8.3575   1.4883   3.7468
#>    8.2677  -2.0748  -0.3816  -5.2321  20.3290  13.5382  -8.7840   8.2436
#>   -2.1281  -7.9031   7.7236   6.7489  -8.0298  15.0070   2.0390  -5.0101
#>   -3.6050  -2.9478  -4.4058   1.1405  26.1479   2.9125  -5.3644   8.3157
#>    5.0475 -13.3496  -7.9177  17.6923   1.3917   4.1184  -0.7301  15.0829
#>   -6.4697   1.4147   6.4580  -6.4392  -0.7604   2.7638   4.0272   1.0139
#>   -6.1255   0.9011  14.5848   6.5701   8.2373 -12.2389   2.7855   5.7370
#>    1.0414   7.7419 -13.5086 -12.9534  -1.4644   1.6666   0.7192 -18.1943
#>    3.4850  -4.3118 -13.4210   8.3404  -9.0981 -10.2628   1.0592  -5.1099
#>   -5.6377  -8.5616   6.5839  -5.3602 -10.9004  -6.1134  14.2148  -0.4657
#>   14.5012  -2.9925  -3.9655   1.1298   2.2056   1.7894  -7.2036  10.3622
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]