Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -7.9998   1.4101  -9.5950   6.9076  -7.1508  -1.8329  -3.3458  -9.1865
#>   -1.6140   2.0881   1.2429   6.5860  14.8845  -4.2657  -4.0518  11.5165
#>    3.9601   0.7151   5.7539  -2.4922  -7.0881   3.7851  17.0946  -3.9621
#>    3.4526  -1.2369  -9.9599  -3.1113   8.9293   0.5976   2.7195 -11.5825
#>    2.4686   0.2318  -5.7144   1.5185  -9.6790   2.7578   6.9818  -5.0352
#>   -7.7333   4.3361  11.6868   6.0150  -4.9877   6.5341 -13.2818 -11.9404
#>   -2.0083   4.9009   8.7409  -6.6990   5.7591   4.0935  -8.2987  -6.6803
#>    5.1574 -10.0239  -2.7184  -8.9709 -15.9918  -6.8434   3.6067   2.9361
#>    3.2688  -4.7615  -7.3767  -6.2406  -9.9733  -1.0341   7.4682  -4.3128
#>   -3.6670   7.0120  -9.4618  -4.6920  -5.6713   7.5760  -7.5893  -4.9214
#>    4.0062  22.2407  -9.5800   4.4215  -2.2242 -17.6076  -9.4431   6.0040
#>   -6.9921   2.3809   1.2586   9.9886   0.8305   8.9747   9.0908  -9.4381
#>   -4.3097 -13.4423  -5.7392  15.2120  -7.7710 -19.5940 -10.1177   2.1235
#>    0.1142   2.7402   1.5522  10.5373  -6.2037   1.1457   2.2351  14.8197
#>   -5.2592  -5.3308   8.9548  -2.3345  -8.2948   6.7082  -7.0611   4.9865
#>   -2.3165   7.8088 -10.3727   6.0061 -15.4599 -16.5779  -5.0688   6.8076
#>   -1.7545  -0.9832   1.2682 -19.9874  -6.3920 -15.7416 -21.0048   3.9997
#>    2.8768  -3.7116  11.0685  -7.4597  -7.8664  -5.2194  -5.5993  -0.3824
#>   -0.1786   2.6688   6.6413   2.3611  -0.5838  -2.7183  -4.3548  15.4481
#>    4.2033 -10.2811  -2.0477  -1.7120  20.2235 -15.1390  20.9091 -15.3033
#>   -0.1251  -7.3943   9.9282   4.0511 -12.1967  19.6561  -7.6562  -3.1273
#>    3.3535  -1.8442  14.2531  -1.9122   1.7578 -13.7271  10.2951 -13.2564
#>    1.1779  -4.7759  -3.1694  11.8659  -4.2980  -2.1334   8.6532   7.0606
#>   -1.7081  -4.4994   5.9728   3.3592  13.2444  -3.4801   0.5181  12.7765
#>   -6.8666   8.2081   3.5271  -1.7829  -7.6765   0.4580   6.0721  -2.2540
#>    1.2864  -7.8262  -5.4305   9.8979   0.6752   3.9568 -11.1754  14.5975
#>   -0.5909  10.3477   4.2768   7.1906  -2.6860   7.2765   0.2332  11.0523
#>    0.9392  17.3670  -2.6578   5.1943   5.1491  -0.7196  10.2995   1.4014
#>   -6.3646   2.0392   6.7197  -7.1569  -2.8875  -0.0765  -0.5248   4.4734
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]