Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.9213  -7.2733   1.3626  -3.3550  -0.6851  -9.1800  -5.8524   8.9731
#>   -0.2102  -3.0091 -10.0165  10.4337 -10.6258  -0.1719 -10.5393  -1.9573
#>    0.0500  -2.7239  -4.2429 -10.9930  -1.0391 -10.0596  -3.2275  -1.6663
#>   -6.2625  -0.8391 -10.5105   2.4530   7.5607   2.1640  12.8285  -3.3405
#>    1.0366  -3.5903   2.2273   7.1928 -17.3678  -7.1638   0.1633   0.0426
#>    2.2927  10.4068  -5.3200  -0.5820   2.8869  -2.7835   1.6747 -16.9459
#>   -0.7187  -2.0355  -3.3506  13.4146 -11.9137 -12.9201  10.3822  13.4286
#>    1.9605   7.4524 -13.6785  -1.9996   5.1061   0.9373  -7.6067  10.4185
#>   -3.0261  -5.8060   2.8961 -10.1343   2.8681  12.9983  -1.3373 -14.8348
#>   -3.8094   1.2805   1.8889 -12.0704 -26.3442  -0.6313  27.8723 -13.4811
#>  -10.2019   8.6939  -4.9040 -13.0921  18.8419   7.8006 -16.4110 -14.1121
#>   -1.9359   0.5578  -4.3392   6.8483  -0.9059   1.9552  -1.3359  10.1124
#>   -2.8709   2.4694  -3.9026  -8.1672   8.3760  -4.2646  15.1444  -7.1268
#>    8.2644  -2.9489  -9.4087   6.3584  -8.0016   2.4984  16.6498  11.0224
#>    5.0017  10.0861  -6.4727   3.1306   4.8754  -2.1871   3.2410  -4.8699
#>    0.0947  -4.7119   1.5823  -5.4760  -2.4466 -14.2622  -5.3553  17.4749
#>   -3.5240  -8.0200  13.9237  -4.8039  -3.2460   5.4146  -9.9932  -0.1622
#>    2.1529  -6.0750   7.8012  -5.4042  -0.8414   0.4849  -8.9848   3.4585
#>   -5.1856   5.4120   3.8356  -2.8575  -3.7839  25.8898 -10.6293  -1.6616
#>   -2.0765   8.7682  -1.6481 -11.3802  29.5343  -4.6621  -6.0449  -2.7088
#>   -6.5591  -7.3912  -2.9660   1.3548   9.9651   7.0727  19.1045  -2.0994
#>   -1.1511   3.4927   0.6229  -1.6695   1.7047   4.6390  -1.0310   5.8009
#>   -1.5455  -0.2387  -6.9798   8.9422  13.0732 -17.5331 -11.0584 -16.8217
#>   -0.4154  -2.4891   4.1924  -3.5223  -1.4694  -8.1833  -1.2525   3.6793
#>    0.9487  -1.2287  11.7025 -14.7010 -11.2646   1.9855  -9.0690  -0.4416
#>    6.3713  -7.2528  -4.0803   8.5295   9.2194  -0.3846  -2.8916   4.4169
#>    6.1666   3.2257   4.2836   8.5153  10.3025   1.7617   4.7442  23.1755
#>    7.9614  10.1024   2.8644   8.8715   9.6912   9.3723  -5.8433   8.7363
#>   -4.0714  -0.0898   7.0086  -5.5301  -5.1427  -2.7578   0.8543  -4.3913
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]