Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.5870  -1.7309  -0.4744  -0.5075 -12.0795  -3.8020 -20.3124  -3.1844
#>   -1.3407   0.8465 -10.9018   8.6201  11.1066  -6.4213 -16.0119 -10.3572
#>   -2.7941  -6.9567  -2.2296  -7.4766 -18.0851  -5.1817 -16.5361   6.5355
#>   -3.8438   5.9021   2.3320  -3.7883   5.2963  10.1139   0.5875  14.9957
#>    4.3672   5.8563  -4.8207  -2.8131  14.9406  -2.4280  15.0131 -10.1351
#>    0.6657  -4.6665   0.3965   9.4064 -23.3997   4.6962  -3.7005  -4.0253
#>   -2.2935   1.3247   0.2816   6.3601  17.2610   5.2302  10.2695 -16.3595
#>    3.9981  -3.7702  -2.6710   4.0293  -5.9239  -1.2572 -18.3067  -5.5621
#>    1.4382  -5.6719   7.6791  -0.4069   4.9455  17.5919  11.2174  19.5007
#>   -0.1602   3.8428   4.0269  10.3329   5.0917   8.9540  -9.3480  -8.9032
#>   -1.9592  -6.6411  -3.1625   2.5375  -7.8072  -5.4790  -1.9595   5.2389
#>   -0.4894  -4.2040   0.7044  -1.5668  13.1444  -7.3873  -6.7423   2.5488
#>   -2.7669  -2.5200  -2.5606  -3.3364   2.2979  -7.4914  10.2009  -4.2631
#>    0.4482   7.5666  -3.6019 -12.6370   2.3844  -3.9560   2.4281  -8.6723
#>    3.0636  -2.5799  -6.6748  17.6716   5.3840 -12.7283   4.1839   0.7627
#>    2.5151   0.3766 -10.9562  -3.4586 -13.1887  11.5112  11.5769  -3.2734
#>    3.3074   2.4340  10.4063  -1.1447 -22.3501   3.1822 -12.8731   1.1578
#>   -3.0651   3.3141  -1.9370  -9.5144 -14.9854  14.2434  10.0303   3.1899
#>    1.1185  -0.6787  11.2806 -15.0825   7.7911   4.4907  -6.7204 -11.5908
#>   -0.5501  -3.6975  17.6507 -11.2861  -9.0992  10.2773   0.0741   4.1716
#>   -2.4953  -5.0628  -7.0247  -0.1722  -6.8803   3.2363   4.2362 -16.2851
#>    6.2498   8.7234   0.0289  -4.7887  -4.2039  -1.9189  -4.1510   5.6241
#>    3.1837   0.9681  -7.2979   4.8138   4.4162   9.0567 -14.0648   5.5451
#>   -0.4644   1.2692  11.1699 -11.9686  -4.3414  -5.3563 -12.7545   0.3729
#>    0.5644  -5.3246  -8.1418  -5.7298  -8.2871  13.2085  -9.8088  -1.4231
#>   -2.1329   3.7523   6.1105  -3.0060  -6.8931   7.6101 -12.5949  16.3728
#>   -5.6037  -6.1351   1.2295 -12.4712   0.3166   1.6718  10.1429  -1.2834
#>    1.9147   5.3166   2.4078  -2.9446 -13.1029   6.0993   4.9132 -12.2409
#>    0.6861   5.4277  -5.7886   5.3328 -18.1003   0.2649   4.7962  -3.7464
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]