Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.1774   5.4386   4.7272   0.9844   0.8426  -5.1550  16.7533   0.7453
#>   -0.0656   2.2226  -3.6709 -11.8241  -8.2700 -10.0253  -0.5007  -7.5811
#>   -2.6899  -0.4881   1.2622  -2.7057  -0.8726   7.5815   4.6585  17.8513
#>    2.8349  -4.4117  -2.7659   5.2571  -8.1009 -13.8393   5.2060  -9.4415
#>    2.3515  -4.5114  -7.6687  -2.8916  -6.6819  22.4954  -7.3245 -15.1091
#>    2.3988   7.8866 -11.5421  -5.1392 -18.2854 -14.4292   2.4762  -8.0510
#>    3.8264   6.3695   0.4377  -4.0888   3.7521  -6.0083  -3.2872   5.5458
#>   -4.0152   4.8601  -8.3256  11.0030   4.7751  -3.3368  -0.8722   3.3962
#>   -4.5096  11.7235   3.6724 -11.3034   9.7259 -17.2987 -20.5907   7.4892
#>   -0.8745 -10.3279  13.7325   6.5277 -19.8222 -13.4315   3.9113   0.4022
#>    5.1340  12.2284  -3.5879 -16.9688   4.6675  -7.7892   7.3696  -9.6347
#>   -0.2647   2.6009   3.3784  -0.6837  -8.5388  10.4899   0.1121  19.9608
#>   -5.6355   3.6903   0.6934  11.7798 -10.9584  -5.7016  -6.2519  11.0117
#>    3.5293   2.3190   1.7541  -3.9933 -11.2297   6.8937  -1.2130  -0.7905
#>    2.9277   6.3725  -9.8290 -17.3973   7.5308  -4.1601  -0.8574  -7.4785
#>    0.1404  -4.3876   3.6862   4.1456  -8.0889   9.4081  -8.6637  -5.5984
#>    0.1088   0.5773  -3.3203 -15.9686  -7.9670  20.7750   2.3016  -7.1624
#>   -0.9774  -6.4594   7.7759   6.2195   4.0081   2.5868   1.9197  -9.3386
#>    3.1042  -0.9075 -10.5381  -3.3261 -17.4515   9.6253  -4.3384  -6.8458
#>   -0.2691  -2.5250   0.7024  -8.2629  11.1995   0.5052 -13.6759   3.4168
#>   10.5519   7.8277   9.3270  -8.7931  -9.4381  10.0274  10.1289   2.7124
#>    2.3882   9.6516   5.9494  -4.9930 -12.0103  -2.1151  -4.9116   7.9952
#>   -3.6098   3.6363   3.2619   1.9068  -4.7503  -2.5071 -24.4788  -3.9836
#>    6.2663   1.4967  -7.3723  -5.2100  -8.7452  -3.7528  -1.3828  -8.1765
#>   -2.1652   3.3912  -7.7849  10.6310   3.7738 -18.1085  -4.1325  -3.7723
#>   -2.4335   9.5103 -14.4860  -8.1174  15.8322  -4.0959   0.9582   3.2979
#>   -0.2665  -4.4231 -12.9690  19.3857   3.4040   8.7117   8.4314  -4.0320
#>    1.9734  -6.0957   7.4836  -7.2188   1.7578   4.9439  10.9020 -11.0034
#>   -0.3617  -3.5939   2.7018  -4.9313 -10.1711 -14.1952  13.3195  -7.0031
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]