Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -5.2964e+00  8.0514e-01  1.1627e+00 -1.9675e+00 -2.2277e+00  2.1469e+00
#>  -1.1567e+00  4.0379e+00  3.7409e+00 -5.2936e+00  4.8562e-01  7.4153e+00
#>   5.6531e+00  4.0550e+00  5.3291e+00 -1.8882e+00 -4.9233e+00 -2.9270e+00
#>   3.0968e+00 -2.0619e+00 -4.7385e+00 -2.4024e+00  6.1223e+00  5.8946e+00
#>  -1.7829e+00  4.3790e+00  6.3045e-01  7.0998e+00 -2.2700e+00  2.0927e+00
#>  -2.5432e+00  4.2571e+00 -2.8729e+00 -9.7644e+00 -5.3124e+00  1.2422e+00
#>   7.0783e+00 -1.0798e+00  1.4330e+00 -2.9545e-01  1.0509e+00 -8.7666e+00
#>   1.2721e+00 -1.1024e+00 -4.8880e+00 -6.8539e+00  2.2045e+00  2.7540e+00
#>  -1.0504e+00  5.0847e+00  4.4767e+00  3.4998e+00  4.3609e+00  2.1574e+00
#>  -2.9258e+00  2.0205e+00 -1.0916e+01 -4.4883e+00 -7.3843e+00  2.6753e+00
#>  -2.2521e+00  1.6844e+00 -1.6912e+01  3.6500e+00 -6.0906e+00  1.0058e+01
#>   9.8410e+00  5.8950e-01 -1.5057e+00  7.0572e+00 -1.3767e+00  7.1931e+00
#>   3.2456e+00  7.5078e-01  8.5479e+00 -4.2698e+00  1.1519e+00 -1.0996e+01
#>   1.6056e+00  2.8321e+00 -1.7063e+00  5.6911e+00 -3.3679e-01 -4.6183e+00
#>  -4.9333e+00 -6.7600e+00  2.5260e+00 -7.1962e+00  1.6294e+01 -9.0772e+00
#>  -6.0054e+00  4.6923e+00 -2.8795e+00  1.8530e+00  1.2211e+01  4.9641e+00
#>   2.7031e+00  7.9413e+00 -1.0900e+01  1.3073e+01  6.8451e+00 -1.4778e+01
#>  -3.9144e+00 -3.8966e+00 -2.0171e+00  1.8189e+00  4.8969e+00  7.9851e+00
#>  -4.2440e+00  9.5377e-01 -6.7005e+00 -6.6536e+00 -5.1970e+00  2.0391e-01
#>  -4.0667e+00  4.1927e-01  1.1093e+00  4.6496e+00  5.1364e+00 -6.0230e-01
#>  -6.5449e-01 -5.7461e+00 -4.4877e+00  1.1044e+00  3.3034e+00 -5.3726e+00
#>  -2.0863e+00 -3.1253e+00 -3.5304e+00  2.6032e+00  9.3716e+00 -8.4192e+00
#>   7.1983e+00  3.6204e+00  5.2014e+00 -6.6063e+00  3.0287e+00  6.3861e+00
#>  -2.8048e+00 -3.8810e+00  1.7541e-01 -1.7961e+00 -7.9508e+00  9.4411e+00
#>  -6.1846e-01  5.4886e+00 -5.1140e-01 -3.8697e+00  6.8423e+00 -1.0243e+01
#>   2.2300e+00  1.4485e+00  2.6612e+00 -1.1104e+01  4.3518e+00  4.0531e+00
#>  -4.5617e+00 -1.4158e-01 -4.2217e+00  1.1349e+01  1.1233e+01  1.0944e+01
#>   3.7688e+00  2.9233e+00 -3.9903e-01 -4.2587e-01  6.9688e-01  1.0142e+01
#>  -2.9410e-01 -3.4495e+00  4.1512e+00 -5.4647e+00 -3.5833e-01  1.5112e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]