Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.2959   5.8666   3.1256  11.0741  -6.3076   6.2992  -4.8116   1.1988
#>    0.0523  -2.1731  -6.6701   4.2152  -0.6517   5.1443  -7.7510   3.0788
#>    2.6034  -1.0530  -3.2724  -4.2813   8.1173   5.2519  13.1131  -6.9176
#>    2.8928  -8.9488   6.6655 -16.9058   8.9674  -0.4090   4.8555  -7.5332
#>   -0.1972  -1.0900 -10.5463   2.3122   0.2225  -1.2950 -11.2001  -4.4835
#>    3.8452  -2.7203  -6.8743   7.6349   4.0599  18.5056  -4.8308   4.0932
#>    2.1249   4.7263  -2.2722  -9.7407 -15.3182  11.2710   1.6543   6.6988
#>    0.7165  -0.3608  -4.1417  -7.8074   9.5905  -3.9740   7.8802  -2.2627
#>    0.4684  -6.3203   5.4986  -7.8241  12.9318   4.1952   5.3181   7.2363
#>    1.0215   7.1993   3.5863  -0.4152  10.0757  -9.3278   4.8915  -2.3958
#>   -6.1184   3.8332  -5.3875   7.1620  -7.0066   0.2890 -14.6752  -1.1263
#>    2.8434   1.6496   5.1703   8.9358  -3.9500   5.4824   2.2737 -12.3147
#>   -0.0576  -0.7158  -2.1239  -0.5288   0.8793  -4.6335   5.1617  -8.0916
#>    4.0354   6.9799  -0.9097  -4.2766  -8.4916   7.3130   4.1399  12.0186
#>   -1.4964  -3.9618  -0.1316  -1.3101  -9.3058  -2.8212   3.5247   3.8440
#>    6.0094  -1.2706   3.0035   4.7164  -9.3794  21.5310   5.3150  13.0267
#>   -0.8621   0.1395  -1.6131  -2.0881  -3.7475  -8.6387  -4.1452  -4.9349
#>    0.3400 -11.6364   5.3610   1.2125  -1.2081   1.0606 -10.5275 -11.6160
#>    0.3241  -1.8833  11.8884   3.7715  -0.6146   4.3725  -3.9664   9.8673
#>   -5.1077  -6.4838  -1.9600  -0.9019  -1.3721  -7.8972  -5.1573 -11.5265
#>    3.0605   0.2665   6.0898  14.6463  -6.5172  11.2039  -5.4553  -8.6491
#>   -2.2566 -12.3601   3.5883  -3.4598  -3.3916   5.4973  -5.5591  11.8615
#>   -4.5361  -5.6819  -4.8503  -5.5180  -2.9258   2.2359 -19.1971  -4.1189
#>   -4.6634  -3.5263   5.5971   0.1566  -2.7550   0.0202 -11.0492   4.2065
#>   -2.8529   1.7602  -1.6359   3.6739   4.3753  -3.6171   5.1127  -5.9433
#>    4.0133   8.4343  -5.2040 -10.3419   3.3185   6.9385  -3.3520  17.5079
#>   -1.5946  -1.8574   7.5401  -8.9921  -2.3857   4.3034  -1.4216   3.8717
#>    6.2910   6.3449   9.3392   0.8800  -2.7912   6.6897  -2.8037  -1.3921
#>    2.4311   4.8676 -13.3440  14.9637 -14.7314   4.8397  -3.7209   0.4603
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]