Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -6.0526  -4.3944  -2.7835   8.1284  22.9698   9.0961  -2.6043   0.7020
#>   -4.6007  -3.5772  -1.6678   0.2785   5.0217  15.2592   6.4545 -11.3940
#>   -1.6703   5.5051 -11.5691  -6.4610  18.2594  18.1424   4.4580  -3.8981
#>    8.7290   2.5702  -8.3349   5.3615  -8.9630   3.7801   4.5015  24.0386
#>   10.0208  15.0870  18.4477   6.7225  -4.6510  -9.2308   1.7872  -9.7117
#>    2.0939   0.2993 -17.6232   6.9101   4.0185   2.8826 -10.4663  -5.0288
#>    0.8788  -5.3259  -1.7810  -7.9740 -10.4938   8.8456  -2.3080   1.5344
#>   -0.1787   6.5390  -1.5690  16.7062   1.1664  -2.5877 -16.5137   8.6250
#>    1.3359  -8.3183 -14.4644  -0.8109  -6.8931  -8.1903   1.7178  -8.3832
#>    3.3576   1.3410   1.5163  -3.7254  -0.7964  -4.9943   8.9084  -8.9685
#>    2.2903  -5.3366 -15.7890  -4.5724  15.2331   0.4391  14.8756 -13.1932
#>   -1.3458   3.1104   2.8397  -2.1662  -7.6187  16.1989  -9.5222  -6.3361
#>    3.7582   3.5365  -8.8219  -6.9294   8.4525  29.9793   3.0377 -14.8064
#>   -3.2306  -5.1290  -6.3375   9.3600  -0.7950  -7.0814  -0.8811   3.6944
#>   -0.8152  -3.9099 -13.5042  -0.2571  -8.5795   6.1095  13.9634   0.3703
#>    6.8096  -2.4500   9.4771  -1.9342  -9.5614  -6.6222 -12.3083   3.9870
#>   -3.8076 -10.5600  -2.0145  -3.2256  -0.6249   9.1207   7.5004  10.8779
#>    1.6452   0.7791   6.7756   9.8536  12.2289  -9.7961   9.3147  -1.6083
#>   -4.6432  -0.2738   6.6062  -6.8122  -6.8501 -11.7113  -0.0598  -4.8704
#>   -0.9291   6.3159  12.3862  -3.0862   7.0742 -10.6970 -14.6802  -3.3646
#>   -4.2904   0.0193  15.6052  -1.1813 -18.1730  -1.2642  -7.2378   2.7837
#>   -2.4532   1.2946   3.5052   3.6402  18.4400   7.1347  -8.1392  -7.5304
#>    3.4441  11.3581  -5.7939  -6.2777  -6.2283  -6.7537  -3.3996  -0.4881
#>   -7.9390   3.3568   3.9860   4.8174   5.6529   2.0800  -2.8714  -4.6299
#>   -2.5962  -8.3158   0.7263  -7.1002   7.8452  13.1240   5.5722 -22.2287
#>    7.4932   6.0281   9.4708  14.2516 -10.4233 -11.6174 -11.4333  -3.3529
#>    7.2956   0.2900   1.1452   0.2760  -3.2748   3.5596 -10.1299   5.5467
#>   -2.9941   9.4619  -4.7313 -16.5435   1.2944   6.3897  -2.1676   9.2862
#>   -1.0936  -2.5538 -12.5369  -9.3060  -3.1822 -11.9254  -4.1164  -7.3039
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]