Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.7592   1.6104   2.0931  -1.0043  -2.6591   3.1942  -7.0184  -7.6719
#>   -1.7238  -2.5490  10.4149 -11.9347   1.1879   3.8831   2.0875  -4.0069
#>   -3.5060   3.7939  -8.3068 -18.4050   2.0903   9.9308  -2.4818  -9.4831
#>    0.9756   6.4314  -4.7985  -4.5925  15.4708  -3.1458  -0.6711  10.0799
#>    2.4947  -0.7026   9.1824  -9.9313 -10.7345   9.0898  14.5556  -4.9983
#>   -3.0655  -3.8055   4.2360  10.6355   2.7476  -4.0014  -6.5690  -5.0191
#>    0.2064  -7.0861   5.9476  -4.5569   7.8547 -23.0478 -11.2328  -9.5004
#>   -1.5481  -1.3110   2.9803  -3.1554  -3.9131   5.8806  18.2287 -18.6942
#>   -2.7719  -4.9730   1.5876   7.2965  -6.2764  -0.6981  -8.9902  -8.0687
#>    3.5723  -0.3914  -0.3429  -0.2727   5.1646   1.7415  -5.9863  -0.7366
#>    1.8378  -4.1656  -6.7451   1.7072  -5.2417   1.2967  -3.9983  -9.0906
#>    3.3126  -6.6392   8.7509   2.2265   1.9872 -11.0770   7.3591   9.6302
#>   -4.0083  -1.5772  12.3707   5.5533  12.9649 -18.1239   1.0019  -2.3526
#>    0.0777   4.0497  -3.7398 -14.8892  -5.2348   1.0817  -1.5106  -2.0587
#>    1.4000  -0.1434  -6.6124  17.3204 -34.2244   9.6611  -5.3670   4.8545
#>   -1.2474   7.8180   2.8822  10.5359  -4.6356   8.7554   1.4678  -2.6608
#>    3.0081  -4.0878  15.2731 -10.7534  -1.0090   9.5663   7.5786  13.8722
#>   -2.6502   3.0164  -0.5715  -9.6637  -8.0663  -1.1294  -5.0861  -0.8285
#>   -6.4548  10.9684  -5.4902   6.6315   7.5372  -3.6023   2.8302   3.8688
#>   -0.8607   8.5697 -12.3460  -6.0350   7.8631   3.0324  -0.0331  -0.2115
#>   -1.0403   5.3189  -4.8519  -3.3070  -5.9236   9.1252  16.6187   6.0132
#>   -4.2472  -2.5522   2.9755   2.3688   1.6718   0.5967   8.5161   5.0279
#>    1.0185   6.3026  -9.1510  -8.7951   2.7732  -2.6839  -0.9554  -1.9799
#>   -9.9197   6.0021   0.8943   3.9916  10.2618  -3.2953   2.3822 -13.6572
#>   -3.0680  -9.4136   7.4962 -11.2653  -3.9720  -1.2654  -2.8529   0.5619
#>   -1.1814  -1.7607   0.5211   4.0798   0.9024  -2.2708   2.4903  12.3212
#>    2.1489  -5.0979   4.0711  -0.5295   0.8323   1.7146 -18.7287  -9.5730
#>    1.9525  -8.4285  -1.7246  -1.1510  -2.9294  -1.9100   0.3139  13.7363
#>    4.2477  -7.3841  -2.1526   3.3101   3.3194  10.2806  -1.5696   0.7466
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]