Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -1.4883e+00  1.2436e+01  4.1880e+00 -4.4747e+00 -5.3737e+00  1.1905e+01
#>   2.1345e+00 -8.0018e-01 -8.5759e+00 -3.9567e+00  5.2878e+00 -6.6629e+00
#>   3.7695e+00 -5.5403e+00 -2.9519e+00  3.0321e+00  4.6778e+00  9.5236e+00
#>   2.8698e-01  1.6078e+00  2.5775e+00  4.0728e+00 -3.9542e-02 -4.6825e+00
#>   1.0057e+00  6.1341e+00  6.1660e+00  3.9254e+00  9.6431e+00 -6.3929e+00
#>   3.3514e+00 -1.5887e+00  4.7865e+00  5.0365e+00 -3.5817e+00  1.3279e+01
#>  -1.4191e+00  2.4775e+00 -1.9759e+01  8.6589e+00  5.0833e+00 -1.1096e+00
#>  -2.8904e+00  9.2828e+00  5.7627e+00 -4.1188e+00 -1.8562e+01  1.3213e+01
#>   4.0680e-01 -4.5701e+00 -5.4894e+00 -6.0654e+00  8.9754e+00 -1.6725e+01
#>   3.1576e+00 -1.0088e+01  2.3395e+00 -3.3183e+00  9.9309e+00 -8.1214e-01
#>   9.5591e-01 -2.4754e+00 -9.3719e+00  5.1243e+00 -6.4453e+00  1.1425e+01
#>   2.4718e+00 -7.6220e+00  1.5773e+00 -1.2484e+00  1.7508e+00 -5.4500e+00
#>  -2.6578e+00 -2.5776e+00  6.8396e+00  4.9203e+00  1.2841e+01  5.9499e-01
#>  -2.9686e+00 -8.4626e+00 -1.8778e+00 -2.8741e+00  1.3142e+01 -8.3080e-01
#>   2.1010e+00 -1.3193e-01 -1.1540e+01 -7.8431e+00 -9.7504e+00  6.3831e+00
#>   1.9634e+00 -2.2276e+00  1.1558e+01  5.9012e+00  1.2731e+01  9.5535e-02
#>  -9.3557e-02 -5.2352e+00 -4.6384e+00  1.3175e+01  1.5970e+01  4.4984e+00
#>  -2.1923e+00  5.4974e+00  9.2770e+00  2.2761e+00 -7.8838e+00  5.2123e-01
#>  -4.3767e+00 -2.1397e+00 -6.5478e+00  1.1326e+01  2.5162e+00  3.4837e-01
#>   3.3919e+00 -4.5573e+00 -1.9311e+00  6.4452e+00 -1.1963e+01 -1.4682e+01
#>   1.9553e+00  7.0567e+00 -3.3084e+00 -7.4924e+00 -3.7740e+00  1.3471e+00
#>  -4.7926e+00 -3.1582e+00 -1.1089e+00 -2.2344e+00  1.5343e+01 -4.1349e+00
#>  -1.5686e+00 -1.0744e+00  4.5874e+00 -5.5205e+00  1.1074e+01  5.0040e+00
#>   2.7850e+00  7.7987e+00 -7.2001e+00 -1.2740e+01  1.4909e+00  1.3146e+01
#>  -4.0152e-02 -5.3561e+00 -1.1366e+01  6.2225e+00  9.7264e+00  5.0298e+00
#>  -1.9107e+00 -2.8802e-01 -9.2359e+00 -4.1925e+00 -3.1952e+00 -2.6149e+00
#>   9.0578e-01 -3.1349e+00 -1.2565e+01 -1.2780e+01  5.7099e+00 -3.4992e-01
#>   2.9657e+00  1.5815e+00 -1.2482e-01 -1.7836e+00 -6.3122e+00 -4.3677e+00
#>  -4.8880e-01  4.8955e+00 -4.9751e+00 -1.4336e+01 -5.6088e+00  1.2520e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]