Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  5.4701e+00 -2.7353e+00  4.2830e+00  1.7378e+01 -7.4390e+00 -9.5727e+00
#>  -4.4132e+00 -6.2669e+00  1.5899e+01  1.6077e+00 -2.3983e+01  9.5867e+00
#>  -4.4756e+00  5.9384e-01  6.5281e+00  8.2736e+00  2.1012e+00  4.5171e+00
#>   5.1199e+00  8.7363e+00  2.3974e-01 -4.4736e+00  2.9563e+00  2.7829e+01
#>  -5.3371e+00  1.2346e+01 -2.1250e+00  1.8102e+01  9.4057e-01  7.2826e+00
#>   1.9766e+00  2.2750e+00 -4.7060e+00  4.8193e+00 -2.5083e+01 -5.4523e+00
#>  -6.2213e+00 -9.3578e+00 -7.2194e+00  4.7773e+00 -1.2084e+01 -3.4433e+00
#>   6.8476e+00 -6.4947e-02  4.6002e+00  2.2000e+01 -4.5395e+00 -7.2362e+00
#>   5.2182e+00  3.7232e+00  1.5927e+01  2.3273e+01 -9.4990e+00 -6.0744e+00
#>   8.6588e+00 -1.3107e+01  5.0582e+00  2.1431e+00 -6.1613e+00 -1.6752e+01
#>  -2.0187e+00  8.0973e+00 -1.1789e+01  1.0353e+01  9.1508e+00  7.1410e+00
#>   1.7756e+00 -2.4798e+00  4.3413e+00 -7.0053e-01  2.2610e-02 -1.2126e+01
#>   3.4680e+00 -2.4217e+00  3.4288e+00  1.2093e+01  7.9038e+00 -7.7807e+00
#>   1.4757e+00  8.5992e-01 -1.1669e-01 -3.9896e+00 -3.1146e+00  4.8144e+00
#>  -1.1763e+01 -1.7373e+00 -9.5603e+00 -3.2053e+00  6.6103e+00 -1.2339e+01
#>   3.7410e+00  1.6830e+00  7.3582e+00  3.4415e+00  3.7607e+00  9.2308e-01
#>   3.4773e+00 -8.3649e+00  6.4866e+00 -3.5690e+00 -8.0573e+00 -1.5228e+00
#>  -4.2447e+00 -8.9232e+00 -2.0891e+00 -1.1062e+01 -4.7408e+00 -1.2964e+01
#>   4.8781e-01 -9.3246e-01  1.0976e+01 -3.0981e+00  2.5569e+00  4.3914e+00
#>   3.0975e+00  9.9757e+00  7.3128e+00  9.6659e-01 -3.9578e-01 -2.9732e+00
#>   1.4311e+00 -3.7661e+00  4.5944e+00 -1.3279e+01  5.3542e+00  1.0994e+01
#>   5.7166e+00 -2.6773e+00 -1.7218e+01 -1.1746e+01  1.6022e+01  4.4765e+00
#>   3.4847e+00  3.2794e-01  7.2563e+00  1.6177e+01  3.3724e+00  1.0178e+00
#>   1.7590e+00 -1.5578e-01 -9.5870e+00 -8.8659e-01  1.7943e+01 -4.5635e+00
#>  -3.1679e+00  1.8720e+00 -5.3859e+00 -2.9033e+00 -8.2784e-01  1.5246e+00
#>   2.3207e-01 -4.6291e+00  1.0205e+01  3.8511e+00  4.5170e+00 -4.6676e-01
#>  -5.8484e+00 -1.2210e+01 -5.6790e+00  3.0513e+00 -7.7334e+00 -1.2991e+01
#>  -1.5742e-01  4.0793e+00 -8.3357e+00 -2.5478e+00  1.4452e+01 -5.8185e+00
#>  -1.6413e-02 -3.4658e-01  9.1415e+00 -9.7883e-01 -8.5256e+00  6.9693e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]