Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.5674   2.1633  -7.7572  -3.4164   8.4504   3.5232   8.7465   4.6849
#>    3.0964   4.1250   6.3972   5.4244 -13.5676  -2.3688   9.2397 -11.0662
#>    5.6807  -0.5053  -1.5211 -12.5855  -2.2988  -1.4580   3.6230   5.7166
#>    5.0238 -11.3694   2.9734   8.9556   5.0875   3.2522  10.2151  11.0314
#>   -3.1623   8.9438  -0.6797  12.8920  15.1110   4.0473  -1.8556  -4.7857
#>    0.4644   7.6599  -9.3975 -13.8544  20.5459  14.0894  -3.0956   6.5492
#>    0.7933   2.7363  -0.3004   7.4768 -11.1516  -4.6611  -5.9396 -14.3478
#>   -1.2405  -4.3067  -4.6225  11.9252   9.5300   9.4471   8.0626  13.5178
#>    4.4433   0.5294   0.9963  16.0316 -12.7412  -2.8124   4.5832  -1.9600
#>    2.9264  -5.5017  -2.0953   4.9435 -16.8175  -0.0260 -17.7063   4.2684
#>    3.0452  -9.1810  -4.0823  -2.1788   3.9463   7.4790  13.5453  10.6512
#>    0.5448   3.0710  -8.7102  -6.6415   9.3213 -18.7383  -3.7039 -10.5812
#>    0.6195  -0.7820  -5.0456  -4.9816  -2.4688   3.6391  -2.9594   0.1132
#>   -1.7720  -1.9860  -0.8086  13.2437   1.6170   0.1103 -19.7414 -10.0123
#>   -2.9224  -3.8877   0.9442   6.2948   8.5901   9.4318  13.3993  -8.7957
#>   -1.1204   2.3602  -2.0687  -0.8381  10.7836  19.2608  -1.1210  -0.3013
#>   -4.7739  -2.9078   3.3035   6.5248   0.0931   4.8931   3.4319  -1.9099
#>   -1.4258  -0.9178   0.5949  -2.1033   0.8233  -0.4562  28.2837  -5.4023
#>  -10.2339   2.8614   6.7048  -6.5077  -3.5123  11.5094 -14.0307  -4.7310
#>   -5.0685 -10.7753  -3.5472  -4.7259   9.6865   2.9773  -1.5788   0.2111
#>   -3.4486   0.8679  -1.0905  -6.4522  -0.8769   4.7247   7.6244  -4.2962
#>   -7.9420  -3.6911   2.5218  -2.6954  14.2369   0.1615  -9.2495  -6.7247
#>    2.8148   3.8105  10.4923  10.5350  -8.0306  -9.9639   6.8683  17.6936
#>   -4.1422  -0.7286   5.3511  10.6696  14.4406  -1.4907   7.3949 -10.3884
#>   -2.1924  -2.2586   3.9177  -2.4231  -4.3863   0.5427   9.4685  12.8595
#>    6.5303  10.9403 -11.6054  -3.2209  -7.1768   3.4504  15.5362  -2.6072
#>   -5.0211  -6.2026  -2.8707   0.8583  14.3135   8.4600  11.5801  15.5192
#>    3.3987   3.2977  -4.7459 -18.0874 -12.2062  -7.9673   8.7870  10.7245
#>   -1.5726  -1.9274   3.0853  12.4730  -2.6891  -0.9353   9.8742 -13.5489
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]