Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -1.2337e+00 -4.9266e+00 -1.6704e+00 -9.9256e-01  6.8077e+00 -6.1858e-01
#>   5.9768e+00  1.9926e-01  1.3595e+00  9.2331e+00  1.4502e+01 -7.4197e+00
#>  -6.7287e-01 -5.7936e-01  1.2221e+01 -1.4388e+00  8.2371e+00 -1.5319e+01
#>  -5.2341e+00 -4.7273e+00 -1.0778e+00 -1.2404e+01 -6.7672e+00  7.5647e+00
#>   8.2616e-01  1.9676e+00  4.0867e+00  7.4542e-01 -8.6457e+00  1.7884e+00
#>  -2.8917e+00  9.4598e+00 -4.2324e+00  5.4354e+00 -1.9257e-01  9.5134e+00
#>   4.1958e+00 -3.0584e+00  2.7350e+00  3.4210e-01 -2.4314e+00 -2.6758e+00
#>   1.6635e+00 -1.1270e+00 -6.0378e+00  7.0018e+00  9.0910e+00 -9.0249e+00
#>  -2.8058e+00  1.5306e+00 -1.7301e-01  2.0658e+00 -4.0342e+00  3.2446e+00
#>   3.6544e+00 -3.6670e+00  2.5389e+00  5.0375e+00  1.0737e+01 -5.5097e+00
#>   1.1182e+00 -3.3459e+00 -4.7592e+00  1.7278e+00  5.0125e+00 -6.3574e+00
#>  -2.7335e+00 -1.1724e+00  1.1486e+00  1.1274e+01 -9.8022e+00 -6.5265e+00
#>  -1.0202e+01 -7.1628e+00  7.1146e-01 -4.3869e+00  4.7645e+00  6.1650e+00
#>  -2.3142e+00  4.9378e+00 -1.2264e+01 -4.1542e+00  6.3922e+00  4.0649e+00
#>  -4.1862e+00  6.5484e+00 -2.8197e+00 -4.5195e+00  9.1820e+00 -2.9280e+00
#>   6.3843e-01  1.1996e+00 -1.4913e+00  8.5581e+00  1.7314e+01 -1.2137e+00
#>  -2.0241e+00  1.2658e+00  7.7915e+00  1.0244e+01 -2.6325e+00  3.1824e+00
#>  -3.7181e+00  8.5519e-01 -2.8051e+00 -6.2956e-01 -1.4211e+00  3.2723e+00
#>   2.4846e+00 -4.2463e+00  9.7580e+00 -7.0237e+00 -4.7039e+00 -8.4570e+00
#>   4.3307e+00 -2.8781e+00  5.5158e+00  5.4757e+00  5.5072e+00  9.0272e+00
#>   3.0737e+00 -6.5387e-01 -1.2050e+00 -2.5848e+00  2.5533e+01 -2.0786e+01
#>   1.5192e+00 -2.2233e+00 -7.5063e+00  9.6953e+00  4.0935e+00 -6.1309e+00
#>   3.9052e+00  7.4959e+00  4.8851e-01 -9.9924e+00 -1.5015e+00 -2.8507e+00
#>   1.8006e+00 -6.5121e+00  3.9557e+00  2.3933e-01 -4.3984e+00 -1.0501e+01
#>  -3.9610e-01 -5.3061e+00  3.7055e+00  1.0496e+01 -3.9091e-01  3.0873e+00
#>   3.0675e+00 -3.5466e+00 -4.4701e+00 -1.3838e+01  1.6859e+00 -7.3690e+00
#>  -6.2561e+00  1.2959e+00  7.8764e+00 -7.8652e-01 -1.9265e+00  1.6493e+01
#>  -1.7820e+00  3.5015e-01  2.3012e+00  7.6453e-01  1.1808e+01  2.9758e+01
#>   4.2038e+00  2.7209e+00  1.1879e+01 -1.3091e+01 -8.3854e-01  7.5315e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]