Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.7028  -2.8907  -6.1519   7.2448   6.6571   9.6871  -1.0071  -7.4660
#>  -11.1252  -0.6434   0.2831 -19.4091   9.4458  -7.8452  -3.5905 -11.5507
#>   -0.5695  -8.4093  18.7060 -10.1950  -0.0906   9.1882   7.3681   9.4796
#>    7.2836  -0.8825   9.5004   0.3866  -0.2266 -18.3492   1.2021   3.8271
#>   -2.8538  -1.7686  -3.5551  -4.5502   1.4129  -2.6490  -7.0933  -0.2814
#>  -12.3035   6.5148   4.6973 -10.7051  21.9653   6.7570   4.4854  -6.7978
#>   -0.7682  -6.4786   3.3802   1.8444   1.6928  -1.4811   4.0194   0.1296
#>    7.1449   3.3160   1.9773 -11.4261  17.0611  -3.0062   6.9493 -16.0809
#>   -1.8239   5.5774  -3.5294   7.4593  -2.9858   5.3093  -4.2892 -15.5774
#>    7.2593   5.3384  -7.0584   3.9322  10.8248   2.5095  -8.6627   3.9774
#>   -2.4757  14.0802  -2.9145  10.8712  -2.6505  -0.5877  -4.0222   0.3595
#>    2.5775  -1.1578 -17.1380  14.6087   6.5012  -0.7930  -7.5898  12.3894
#>    4.2505   7.4910 -19.1720 -13.6408  19.0861  -3.3965   2.5777  -4.0215
#>    9.3508  -1.4545  17.2898  -0.8954 -22.2917 -11.4339  -3.6974   2.8669
#>    3.5969   1.0608 -17.9493   4.3394   5.6990  -9.3526  -8.2083  -5.7568
#>   -4.7841  -2.0631  12.5723   3.2796   3.5961  -3.1876  -3.1693  13.0580
#>    1.1511  -3.3075  -3.1462   3.3967  -4.7504   8.4180 -15.5083   1.4060
#>   -0.1077   1.1017  -7.3586  -6.5997  10.0827   0.4689  -4.0522   2.3373
#>    4.9671  -1.7545   8.6951   2.7353   2.6022  -0.5307  -3.2366   4.9237
#>   -3.9535  -3.1572   8.5848  -6.5468   7.3095   1.3120  -6.1009  -3.2008
#>   15.6339   0.0222  -1.3585 -12.4048  14.2453  11.6161   6.7432   9.0397
#>    3.8814   9.3811   8.6933  -3.6201  10.3639   1.1462   4.1630   8.0301
#>   -5.4637 -12.3914  -6.2383  -8.5957 -16.6621  -0.0881  -0.2001 -10.8328
#>   11.2702  -1.4240   8.9038  10.8380   8.5714  -6.1482  -6.3557  -4.6975
#>   -2.7892   3.4092   5.9578   6.3044  -2.9758   8.8815   5.2581  -1.8334
#>   -3.3698  -3.1196  -1.5304  -3.0917  10.9798 -14.7734   4.1340  -8.9719
#>   -1.9281   7.6226   6.8088  -0.9843  -0.7061  -2.5639  -7.8339  -6.3027
#>    1.2196  -1.6766  13.6601 -18.6460  15.6395 -12.1546   5.3204   2.3869
#>   -5.2087  -4.8184   1.9796   2.2806   2.7155 -17.9141  15.4551   0.3079
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]