Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -1.7200e-01  1.7848e+00  6.1791e+00 -1.9975e-01  3.6703e+00 -9.5042e+00
#>   1.5793e-02  5.0277e+00  2.3760e+00  7.4134e-02  2.4026e+00 -4.0351e+00
#>  -1.3413e+00  1.2357e+00 -3.0502e+00 -1.0933e+00 -1.4421e+01  8.8722e+00
#>  -4.5010e+00  7.9632e-01  8.2635e+00  2.6164e+00 -1.2564e+01  1.1762e+01
#>   3.2443e+00  1.4271e+00  4.2374e+00 -5.5736e+00 -1.6198e-03 -6.7487e+00
#>  -5.5031e+00  7.5043e-01  1.1777e+01  2.4440e+00  1.2352e+00 -8.5297e+00
#>  -4.7365e+00 -7.5594e+00  9.2945e+00 -2.5133e+00 -4.8731e+00  1.1365e+01
#>  -3.6561e+00  7.0458e+00  4.0253e+00 -9.6349e-01  5.5444e+00 -2.5490e+00
#>   2.3627e+00 -1.4558e+00 -5.5811e+00 -4.8452e-02  1.5316e+01  4.9384e+00
#>  -5.1307e+00  1.3362e+01  6.3237e-02  3.2101e+00 -3.9881e+00  1.3590e+00
#>   4.8691e+00  1.1134e+01 -6.0463e+00  4.6285e+00 -8.8477e+00 -1.0951e+01
#>  -1.8802e+00  7.4189e-01  8.1038e-01 -1.7731e+00 -3.0575e+00  4.2995e+00
#>   3.7047e+00 -2.0935e+00 -6.1170e+00 -3.0017e+00 -2.8256e+00 -5.6616e+00
#>   1.3730e+00  6.0980e+00 -2.7578e+00 -9.3142e+00 -1.4464e+00 -6.5865e+00
#>  -5.5954e+00 -5.1989e+00 -8.2407e+00  2.9000e-01  2.8342e-01 -5.3953e+00
#>  -1.3028e+00  7.8664e-01 -4.3427e+00  2.8635e+00 -1.1204e+01  7.7065e+00
#>  -6.0062e+00 -4.4630e+00 -1.7183e+01  1.7191e+01  3.6053e+00  8.6722e+00
#>  -2.3179e+00  2.1281e+00  1.7186e+01  4.2373e+00  6.6954e+00  8.6691e+00
#>  -3.3201e+00  3.1044e-01  6.7949e+00  8.1578e+00  7.8915e+00  4.4185e+00
#>  -2.4637e+00  5.1599e-01  1.4021e+00  4.0327e+00  2.5022e+00  5.0723e+00
#>  -3.4041e+00  1.0636e-01  9.5211e+00 -5.5493e+00  8.0142e+00 -6.8941e+00
#>   6.1751e+00  3.3478e+00  4.8102e+00  1.0419e+01 -5.8418e+00 -5.8425e+00
#>  -1.6830e+00 -1.4885e+00  1.3611e+01 -6.2680e+00 -7.4129e+00 -9.9254e+00
#>  -1.4787e+00  3.2000e-01  3.6537e+00  6.5885e-01  1.7984e+00  7.1168e+00
#>  -3.5506e+00 -5.0578e+00  5.9073e+00  1.4700e+00 -1.2740e+00 -6.2489e-01
#>   2.2151e+00  4.3465e+00 -9.8846e+00 -6.9853e+00 -5.4778e+00  1.8094e+00
#>  -6.5322e+00 -5.0052e+00  2.8314e+00  2.8825e+00 -3.6691e+00 -8.4774e+00
#>  -2.5092e+00  2.5202e+00  1.0620e+01 -1.3319e+01  2.0308e+01  4.1457e+00
#>   1.9178e+00  3.1676e+00  1.6282e+00 -2.3304e+00 -1.1024e+01 -1.4104e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]