Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.3612  -6.1165   6.6636   8.6205  -3.4140  12.3388  -1.5479  -5.1406
#>   -5.8006  -5.6633   1.3297  20.3841  -1.8962 -12.9657  -3.3472   2.1724
#>    1.7923   3.1274   9.3181  -3.0663  -0.0661  -6.6948  19.4713  -4.0729
#>    7.5638   9.8081  -4.7958  -7.3679  -1.9487   5.7758   6.4381  -1.7130
#>   -5.8675  11.6418  -1.0948   4.5029  -7.1343  12.5768   3.7000 -17.5119
#>    2.6242  -0.8411   0.8630   0.7156  11.3658  -5.0104   4.2807   6.6471
#>   -0.3196  -3.4966   4.9106   6.8765 -12.2823  -1.5994  -9.1579  17.7404
#>    2.1238  -2.5809  -6.4184   2.0267   3.6044  -3.0641 -12.5162   1.4693
#>   -3.2789  -0.4169   0.2507   9.2139  12.1759 -12.8329   9.9174   4.5352
#>    0.8196  -4.1471  -5.1255 -14.1391  13.9181 -12.7870   1.7688   0.1079
#>    1.6096   0.0088   3.5917  -8.8364   5.2629 -10.9089  -4.0886  -5.1408
#>    1.7713   5.8799   1.4723   2.3564  10.9448   4.0176  -6.8718 -12.4196
#>   -2.5636   3.1872 -10.9821 -15.0678  -3.1090 -13.8802  13.9801 -10.1404
#>    6.3213   7.3165   2.3752  -0.9917   5.0960   0.8876   3.8910   8.2202
#>    6.2525  -4.8684  -9.0993  -6.2851  -9.5597   0.9025  11.4119 -15.7572
#>   -0.8176   1.5159  -1.2490  -5.7181   4.8608   2.2946  -5.6925  12.1223
#>   -4.2415 -15.4087  -0.8608   2.3660  -3.3076 -14.1787   6.6158  -0.1080
#>   -0.7494  -6.3085   3.3112  -2.0911  -0.7492  -1.7961   2.6907  14.5234
#>   -9.7414  -6.9719   4.9673  -0.0242  -8.6986  -2.0565  -7.0299  -3.0005
#>   -0.3055  -1.3741  -0.4060   6.2419   7.0127 -12.2303   1.4298 -14.5712
#>    0.5447  -3.6723   1.4641   3.7581  14.1131   1.3282 -18.7255  -4.6579
#>   -4.1565  -8.7389  -0.9131  -5.4375   7.4548 -18.8530   9.4490  -4.2751
#>   -9.7139  -5.9420   4.2784   8.9130  12.7983   7.2889   7.5140   7.2496
#>    4.6472  -0.6661  14.1137   6.9027   9.5312  13.9868   3.2858   4.6714
#>    4.0019  -8.6654 -16.5855  -8.4247 -15.1716  -6.4166  -5.1052   3.2522
#>   -4.6086  13.1224  -4.8176   7.0213   4.8547   5.1395   6.5610 -10.3774
#>    2.4034   2.7611  10.9153   0.5629  -4.7891  13.5404 -11.4925 -12.4546
#>   -1.7564  -4.6217  -0.9816  -3.5983  12.9152 -20.9415 -14.4753  -8.2617
#>   -0.2413   2.7610   8.3445  -8.8839  -1.5509 -12.0491   0.4816   5.8984
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]