Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   8.0247  -2.3924  -6.0450   7.0425   5.3296   8.9916  -1.1724   4.6660
#>    4.0721  -8.5381   1.7916  -2.1767  -5.3869   5.4069  13.7724   0.6381
#>    0.2591   5.5348   1.4387   4.2608  -8.7481  -6.4023  -3.2975   0.2171
#>   -3.2384  -0.6002  10.5636   9.1190  -1.3838 -13.4011  -7.7066   1.2203
#>   11.8716  -0.9213  -0.7124  -5.7737  19.1555   5.6579  11.6576  -2.6160
#>    3.3451  -3.9594   8.4314   4.9223   7.6233   6.3811  11.1689  14.5655
#>   -5.9365   0.0982  -0.3880  -6.7220  -8.9578   4.1187   2.6446   7.6831
#>   -5.8431   5.2895 -11.2201   5.3057 -16.4976 -21.9900   9.7430  -6.9899
#>   -5.1245  -2.7890  10.7997   6.9912   3.0140  10.4780   0.4029  -3.3847
#>    0.8443   4.3109   1.4129  -2.0966  10.5769   5.7572   4.8961  -5.6739
#>   -3.1305  -0.2020   2.9875  10.3480   3.4470 -10.2043  -9.1976  -7.4600
#>    1.1232  -6.4699   6.9449  -2.5497  10.3091   0.1445  -1.4349  -3.7588
#>    0.7535   0.0355   2.1813   2.8627  -7.7643 -14.7878   4.2790  22.4981
#>    5.3968  -4.9736   4.7038   6.2087  12.0663   4.8191 -10.1318   0.5431
#>   -5.2962  -2.8299  11.4315   7.3615   4.5287   0.0237  -0.4807  -4.1217
#>    6.3993  10.0778   4.8318  -5.9623  -5.6126  -3.0177  -2.5085  -6.0044
#>   -5.4614 -11.5398   7.6094   5.4064  -2.8457  15.0379 -12.4093   0.3664
#>    2.4580   7.0874   4.8669   2.9694   3.0345  -9.1142   1.3417  20.1104
#>    7.7325  14.7454  -4.3295   2.0012  11.5659  15.8893   3.5685   6.1133
#>    0.5918   6.1335  -9.8223  -1.1059   6.4462   8.1892   2.7047   5.4258
#>    4.7170  -3.7191   0.3378   5.1216  -2.1955   5.9489   6.2064   4.9865
#>   -3.2376   1.5301   4.3549   5.3799   5.0293   7.5611  -1.0653  -2.2861
#>    4.4577   2.1079  -0.8596  15.7452   1.5904  12.7106 -10.3194 -14.1728
#>    0.2357  -2.3855  -9.0400   4.0323   7.7257  -9.1412   2.2689   5.5637
#>   -0.4714  -1.9014  -1.4706   0.6587   6.4632  -3.8134 -10.2406  20.5641
#>    3.2587  -4.5678  -8.5975   8.8055   1.9202 -16.9276   5.6236  -8.9676
#>    7.2904  -0.3134  -4.1859   5.2487  -0.2128   5.2203   8.2912  10.2415
#>    3.9877   2.6735   3.0651   9.3555   1.0493  -4.4115   1.3644   5.5200
#>   -0.0128   2.2786   3.2314   5.7658  -9.4384   8.9296   4.3136  -5.2176
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]