Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.3103  -1.8627   6.0041   6.8943  -4.0041   4.7137   4.1134  -2.1084
#>    3.5167  13.7407   6.5199  -1.4415   8.5576  -0.6274   5.1885  12.7256
#>   -3.2110  -6.8645  -9.6220   0.5123 -10.6734  -4.3277   8.3322  -2.6489
#>    4.4287  -1.7557   7.5087  -0.3630   1.5425  16.6489  16.8390  -1.8894
#>   -0.6486   5.5400  -3.1867   3.4929  -1.7472  12.4465   6.1466  14.8881
#>   -2.5059  -1.8231   0.8949   4.7157   4.5559   9.9384   5.3936  -7.8133
#>   13.3650  -3.3263   6.5856  -4.9136 -13.0929  -0.3550  -3.1319 -14.9495
#>   -2.3989  -4.5299  -2.7632  -5.9780  -3.2882   3.3265   3.1559 -22.8117
#>   -3.8384 -12.4963   4.7014  -3.7503   2.2181   5.1535   7.4139  24.2949
#>   -0.8671   0.0017  -3.8780  -4.1856   4.1555   2.3999  -8.3948  -6.4868
#>    0.7823  -7.8695  -0.1304 -22.2776   4.5756 -13.1303  -3.3723  -5.5827
#>   -9.0630   1.7774  -4.3529  -6.1262 -11.6386  -4.4135 -10.1137   4.8471
#>    2.1306  -5.1331   1.4460 -12.4632   1.8560  11.4040  -3.0250 -36.6421
#>    8.2113   1.9318  -3.5601  -2.1804   6.2181  -8.1261 -16.6587   8.9012
#>   -6.4327   1.9328  -4.7717  -2.8566   2.9128  18.6470  14.5958  -5.6198
#>    1.0674  -1.8749  -7.7291  -3.0653 -16.6678  14.1000   5.6421  -1.4141
#>    4.1571  -5.5526   6.7604  -7.5370 -12.5158   4.9489  -3.4926  21.4391
#>    3.5999   0.6997   7.9507  -0.5439   3.0945 -18.8875  -0.5202  -2.5635
#>   -0.3400   1.8249  -7.8909   2.4023   0.2949   1.5831   0.8558  11.7320
#>   -0.0630  -0.1291  -5.0755 -10.8061  -2.6610  15.5673   7.0291  11.2653
#>   -0.2868   4.8493   8.1249 -15.0964  -3.0787  12.2657  18.7564   5.6361
#>    3.1840  -2.6225   3.7090  -4.9759 -18.3992  11.0665  -8.5702  -9.3847
#>    0.0656  13.3279  -8.8639 -15.8049 -13.3920  -4.1100  -1.6985   2.7170
#>    6.1616  -0.7575   0.2055  -9.7888  -5.6569  13.5709   3.3548  -3.5310
#>   -7.4138  -3.4148  -6.9714   1.1604   0.0211   9.5376  -6.2144  -8.4612
#>   -6.6970  -5.7625   1.9173  -3.1688  11.4531  26.2362 -10.6528   1.2697
#>   -5.3310  -1.2341   1.3699  -2.2919 -10.3891 -17.5742  -7.3440   3.8233
#>    0.0168   3.1264   1.0921   2.9197   3.6415   1.3424   2.7723  -0.1677
#>    5.0622 -12.1055  -2.9501 -12.4725  -3.8575 -12.4050  -2.1240   2.3402
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]