Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.8707  -2.1885   0.6030 -10.2505 -11.3125   6.7081   2.8033  -0.8226
#>   -1.0620  -4.3588  -3.6755  17.1357  10.3041  18.6412   1.8714   8.4930
#>   -0.2158   7.3695 -14.4200   5.6917 -15.7038   2.4918  -1.5374   5.3627
#>    1.3176   4.7921  -1.7120  -7.8204   7.9495   3.2523  -5.2002  10.8814
#>   -5.2056   6.7418   0.0028  -1.5313   7.7154   8.4509   1.4129   0.7765
#>   -2.8184  -3.9263   8.0475  -0.5696 -25.0893   3.2822   9.9441  -2.3659
#>   -0.9105  -0.6440   4.1328   4.3684   1.7989   3.2470 -14.5623   6.6662
#>   -9.0706   7.5233 -13.3850  13.5380 -10.7452   5.4740  -1.2713   0.8096
#>   -2.9896   0.4346  -3.6804  -0.4557   5.3363   8.4972   0.0394   8.3137
#>    3.1802   6.8408  -3.5596   1.4764   5.5752  -3.4091   8.4268  -9.1758
#>   -2.1429   7.2652  -7.5654   2.0673   1.5139   9.0099   8.8055 -12.0956
#>    0.6325  -6.1208   0.5467  -6.0638  12.0946   2.4996  -0.0708  -5.4004
#>    0.7876   6.1797  -4.4714  -1.4961   7.9689  15.7256   1.0645  -0.9208
#>   -4.6600   8.3723   0.1464  10.4852  -8.1848   1.1521  -4.1062  11.3779
#>    3.5635  -1.5582  -4.1136   7.1326   1.3673   5.9225   6.8022   6.2023
#>   -2.4765   6.4277  -6.0039   9.2718  -2.9250  -4.4051   5.4509  10.8571
#>   -8.0790   0.7315   7.9564  11.9173   5.3503  -6.4382  -0.4544  -2.0567
#>   -2.7719  -5.8801  11.3601  -5.6794   7.1856  -9.1824  -0.7594   0.7307
#>   -0.3541   4.2886  -6.1563  -5.9588  13.8317  -7.1301 -17.3006   3.9110
#>   -1.5529  -2.5534   1.5002   4.7720  -1.3097 -13.0422   4.9353  -8.3387
#>   -0.1108   2.3689   2.7208 -12.4137   2.9408   8.5481  -3.0780   8.1256
#>   -6.6430   2.3436   7.6377  -8.0378  15.9150   3.7025   8.5762   5.8839
#>   -4.1801   0.5790   5.4706 -11.3501  16.0173  -7.0272  21.2336  -5.1512
#>    2.6364   4.1829  -4.0479  -3.5950  -0.7133  -5.8741  -2.7134   2.2566
#>   -0.8905  -2.8686  -0.3952  10.9920  -4.4680  -0.4352  -3.8917   1.9180
#>    5.6544  -6.1178  -5.2508  10.6930  -1.8400  -7.6333   1.4315   5.7815
#>    3.6335  -2.6110  -2.1697   1.4925  -9.9598 -11.0829  -6.1487   5.4580
#>    1.8030   3.8254  -3.7279  -1.0454 -21.3014   4.3516   8.8147   0.5426
#>   -1.1379  -3.9704  -0.5977  -0.1576  -8.8639  10.4381  -1.7001  -1.3302
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]