Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  4.9968e+00 -7.5526e+00  2.3107e+00  1.9762e+01  8.6400e+00 -2.4185e+01
#>   2.7077e-01  5.7396e+00  1.8254e+01 -8.1642e+00 -3.0417e+00 -4.1452e+00
#>  -1.9622e+00 -9.6577e-01 -1.1503e+01 -1.1439e+01 -1.2180e+01  4.6113e+00
#>   1.8579e+00  1.5017e+00  6.8357e-01 -1.3549e+01 -4.2231e+00  5.6950e+00
#>   4.5328e-01  1.7603e+00 -2.3009e+01 -1.3122e+01 -1.4945e+00 -6.0458e+00
#>   1.1860e+00 -2.6174e+00 -7.2091e+00  7.2548e+00  5.5508e+00 -1.0961e+01
#>  -2.7987e+00 -3.1277e+00  9.3441e+00 -1.3669e+01  8.5802e+00 -1.2941e+00
#>   2.5212e+00 -6.0849e+00 -1.1603e+01 -6.0426e-01  1.1340e+00 -6.3936e+00
#>  -4.7733e+00 -9.7918e+00  2.8373e+00  8.0049e+00 -9.7717e+00 -2.5781e+00
#>  -6.3556e+00 -1.7226e+01 -2.0460e+01 -1.5024e+01 -3.5162e+00  5.1158e+00
#>   2.8971e+00 -1.6215e+01 -9.9241e+00 -2.5101e+00 -2.5718e+00 -5.2292e+00
#>  -8.7608e-01 -3.8312e+00  7.0491e+00 -1.5654e+00 -2.1718e+00  5.5515e+00
#>   1.6111e+00 -1.8845e+00  1.1042e+00  7.4902e+00 -8.1487e+00  4.6557e-01
#>   6.0515e+00 -1.1245e+01  7.6811e+00 -1.2271e+00 -1.3802e+00  1.8219e+01
#>   9.2179e+00 -2.8691e+00  9.5021e-01 -3.9107e+00 -1.6179e+01 -8.6760e+00
#>   6.6124e+00 -2.7487e+00  8.1763e+00  1.0102e+01 -7.2826e+00  1.2932e+00
#>  -2.3399e+00 -6.3203e+00  2.1719e+00 -9.7549e+00  9.4910e+00  2.1748e+01
#>  -5.5790e+00 -3.2488e+00  8.1042e+00 -4.6623e-01  6.1140e+00 -2.2965e-01
#>  -2.9570e+00  4.0963e+00  5.8133e+00  5.3473e+00 -1.1891e+01  1.5405e+01
#>   3.1561e+00 -8.6224e+00  1.2212e-01 -2.5870e+00  2.3039e+00 -2.5492e+00
#>  -6.3256e-01 -3.5638e+00  3.7027e+00  1.1830e+01  1.1562e+01 -3.5759e+01
#>   2.7627e+00  8.6860e+00 -4.6160e+00  7.2252e-01 -7.9443e+00 -2.2446e+01
#>   3.2392e+00 -2.2144e-01  1.4049e+01 -5.6389e+00  9.6289e+00  1.3900e+01
#>  -1.2198e+00 -1.0163e+01 -1.5279e+01  1.0458e+01 -5.9086e+00  1.0229e-01
#>   3.9589e+00  3.7185e+00  6.0094e+00  9.7099e+00  1.0343e+01 -8.8972e+00
#>  -8.1815e+00 -5.4634e+00 -7.5115e+00 -1.9896e-01  1.2370e+01  4.9508e+00
#>  -1.5639e+00 -8.0708e-01  3.2795e+00  3.4012e+00  2.1064e+01  1.0312e+01
#>   1.5933e+00  3.5931e+00 -2.7889e+00  5.4091e+00 -7.6034e+00  8.2775e-01
#>   1.2990e+00 -2.7000e+00  1.2013e+01  1.4618e+00  7.4216e+00  1.0935e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]