Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.1241   6.2136   0.7847   4.8078   3.0414   3.7078   7.2429 -13.9218
#>    6.7129   0.1922  -1.9290 -14.1997  -5.8148  -8.8177   1.8421  -0.1732
#>    7.3974   7.9141   6.0256  -3.3267  -7.5596 -14.0668   9.5888  12.3531
#>    7.8971  -0.3647   7.9184  -0.8190 -14.2175  17.4929   4.7107 -10.1260
#>    2.1095  -1.3399  -7.5166   0.6483   1.6260   3.9077   1.1682 -12.7049
#>    5.0569   5.3562  -3.4514   9.1224  -5.2041  12.2127 -17.3525   3.5195
#>   -5.5517   7.9661  -3.7427   7.9950  -6.9864  12.2143   6.4730   8.9978
#>   -0.4042   1.2245 -10.1081  -9.7187  -6.8353   5.3550  11.1535   4.1948
#>    1.1225   7.9632  -0.6445  10.5199   2.0108   8.0073  -8.7575  -9.4709
#>    6.3725   3.8230   0.4788   4.4231   7.0885 -13.5014   1.0882   1.2434
#>    1.4899   8.8839  12.3112  10.0469  -5.9425 -11.7330   5.0892  14.5632
#>   -0.3098   0.7165   5.8816  -0.4863   2.0227   5.3014   4.0512  10.9960
#>    7.3960   5.4190  -2.6634   9.5366 -13.2210  -5.8485   0.3004   6.1520
#>    0.3877  -3.6698  -1.4110  10.6635  -4.1945  -3.1709   5.9413 -10.3066
#>   -7.4050   7.5717  -1.6772   6.2073   5.4311  -4.4713  -3.2570  -1.1270
#>   -0.3611   1.5048  -2.9581  -6.0500   3.9651   9.9103 -10.1012   0.6746
#>    2.0775   2.8757   0.1023  -2.9862  -1.9776   7.9087  18.4459   4.8111
#>   -0.9553  -0.1785   2.5814   7.0047  -6.9098   0.1241  -1.1928   4.2275
#>   -0.8539   2.6635 -10.6510   3.6605   6.0547  -8.8295  -4.5364  -7.4138
#>    0.6643  10.1050  -2.9439  11.1873  -3.7192   4.3384   9.5737  14.9358
#>    2.8364   2.7882  -1.8958  -8.3092  -3.9609   4.4653   6.5705  -3.7642
#>    3.4479  -4.2742  -8.0148   5.3923   1.7069   6.0466  14.5276  -9.5794
#>   -1.4316  -4.9962  -4.1255  -6.6788  -2.6200  -4.4068  -8.9813   0.2664
#>   -0.2927  -9.0626   0.8289 -10.3048   7.3004   3.0647  -9.0063  -1.5802
#>    1.9008   2.6960  -6.9750  11.9206   1.4900   0.1753  -0.8948  -3.6434
#>    4.4138   1.9646   5.8501   1.8435  -1.7060   2.4465 -11.5927  -3.6810
#>   -0.8375   0.2235  -3.6453   1.5740   9.5716  -2.4541   1.4215   0.0557
#>   -3.4708 -11.9191   5.0340  11.7218  -1.2449   3.7603   5.9853   5.1482
#>   -5.4622  -6.2304  -4.5202  -2.0792   6.5403   6.3345   8.4733   6.3822
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]