Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -2.4138e-01 -2.5180e+00  8.7499e+00  7.5469e-01  1.1294e+01  7.0051e+00
#>  -4.3524e+00 -8.6206e+00  5.5580e+00  1.5746e+01  2.3095e-01 -6.8345e+00
#>   2.9224e+00 -2.8601e+00 -8.8575e+00 -2.5261e+00  1.3932e+01  1.4891e+01
#>   4.4591e+00  8.3416e+00 -2.5294e-01  2.3246e+00  1.9521e+01  1.5089e+01
#>   1.0111e+01  6.3684e-01  2.4004e+00 -3.5228e+00 -2.6207e+00 -5.4745e-01
#>  -3.1271e+00 -5.5721e+00 -2.8641e+00 -1.6909e+01 -5.6417e+00  3.9962e-02
#>   8.5698e-01  4.3989e+00  1.1804e+01 -1.0480e+01 -4.9792e+00  1.3171e+00
#>  -6.8780e+00 -4.5869e+00  6.9407e+00  6.8297e+00  8.0465e+00 -9.1785e+00
#>   9.9326e+00  9.8344e+00  5.9118e+00  3.1602e+00  1.6289e+01 -4.1712e+00
#>  -9.5461e+00 -1.2626e+01 -6.1227e-01 -1.5700e+01  1.7553e+01 -7.4112e+00
#>  -1.6773e+00  4.1468e+00  2.0549e+00 -3.6129e+00 -3.1676e+01  5.1089e+00
#>   8.8566e-01  2.4050e+00 -1.1941e+01  2.0585e+00  1.9654e+01 -2.4406e+00
#>  -4.9673e+00 -1.3440e+01 -4.9772e+00 -6.2764e-01 -1.6405e+01 -9.0714e+00
#>  -5.7636e+00 -8.9300e+00  3.3555e+00 -2.7056e+00  3.5012e-01  1.3860e+01
#>  -3.1756e+00 -3.7669e+00 -6.3029e+00 -1.3919e+01 -1.0427e+01  1.0498e+01
#>   2.8375e+00  1.6162e+01  6.4524e+00  5.8383e+00 -1.0252e+01 -4.6061e+00
#>  -3.0915e+00  6.4110e+00 -5.6597e+00 -4.8491e+00  2.9617e+00  1.3563e+01
#>   4.2670e+00  1.7327e+00 -8.8280e-01 -5.0027e+00 -1.5171e+01 -4.3308e+00
#>   6.2171e+00  9.6955e-01 -5.0446e+00 -3.8485e+00  5.8667e+00 -4.8332e+00
#>  -9.2783e+00 -9.7079e-01 -4.2073e+00 -3.7297e+00 -6.3739e+00  1.2445e+01
#>   2.2750e+00  9.3303e+00  1.5229e+00  7.4488e-01 -7.7927e+00 -4.9781e+00
#>  -2.2949e+00  4.9375e+00  5.7045e+00  5.5162e-01 -6.9700e+00  8.8866e+00
#>   2.8002e+00  8.3107e+00 -1.0834e+01 -3.3446e+00  4.3309e+00  2.5442e+01
#>  -1.2135e+00 -2.9991e-01 -2.1439e+00 -7.9954e+00  4.3624e-01  1.4934e+01
#>  -2.1703e+00 -6.4651e+00  2.0344e-01  6.5230e+00 -3.2017e+00  1.0377e+01
#>  -8.1558e+00 -1.1690e+01 -5.2701e+00 -3.9577e-01  5.5897e+00 -7.3908e+00
#>  -3.8181e+00  5.9279e+00  3.1272e+00 -7.2419e+00 -5.2937e+00  4.9243e+00
#>  -4.4433e-01 -1.1637e+01 -5.1614e+00  3.8272e+00  1.2143e+01 -6.1220e+00
#>  -7.7069e+00 -7.9846e+00 -3.2979e-01  7.4178e+00  5.1651e-01 -1.1158e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]