Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  1.0844e+00 -1.4078e+01  2.1573e-01  5.0276e-01  2.8474e+00 -8.7577e+00
#>  -2.9935e-01  2.5470e+00  8.5472e+00  1.5952e+01 -4.4282e+00 -8.2839e+00
#>   3.8671e-01 -2.7460e+00  1.8398e+00  7.4889e-01  5.2690e+00  8.9464e-01
#>  -6.4981e+00  8.0596e+00 -2.3751e+00 -6.0953e+00 -1.2121e+01 -5.4137e+00
#>  -4.6465e-01 -2.1660e+00 -1.0300e+01  8.3822e+00 -9.3851e+00  8.0734e+00
#>  -3.3277e+00 -3.5608e+00 -8.8888e+00  5.1319e+00  8.5328e+00 -1.6501e+01
#>  -1.6935e+00  1.5141e+00  1.1320e+00 -4.7070e+00 -1.6700e+01  1.0662e+01
#>  -7.2303e+00 -2.7572e+00  3.6541e+00  3.7380e+00 -1.0411e+01 -1.0537e+01
#>   5.1965e+00 -1.6786e+00 -1.1734e+00 -2.5682e+00  7.8338e+00 -1.9477e+00
#>  -1.1218e-01 -5.5703e+00  2.5533e+00 -6.0102e+00 -2.5173e+00  4.0648e+00
#>   3.6351e+00 -1.9909e+00  1.4804e+00 -1.3151e+00  1.3068e-01  1.6822e+00
#>   1.0801e+00 -7.6671e-01  1.9034e+00 -7.1683e+00  3.2826e+00  3.1670e+00
#>   1.7675e+00 -4.0085e+00 -7.9310e+00  7.9552e-01 -1.2682e+00 -2.6007e+00
#>   6.0742e+00 -3.5418e+00  3.6460e+00 -1.4211e+01  5.5859e+00  6.2972e+00
#>   3.9469e+00  3.6522e+00  4.0906e-01 -8.8546e+00 -7.7982e+00 -2.9208e+00
#>  -2.7536e+00  2.5120e-01 -3.7720e+00  2.5303e+00  6.4454e+00  1.3297e+01
#>  -1.7087e+00  5.5031e+00 -4.0421e-01 -6.7333e+00 -8.8779e+00 -5.4687e+00
#>   2.5809e+00 -4.6426e+00  1.6925e+00  3.6016e+00 -7.1125e+00  1.0171e+01
#>   2.1173e+00  7.1775e+00  5.4277e+00 -4.5332e+00 -4.7924e+00 -1.3483e+00
#>   1.3287e+00  1.7026e+00  5.9750e+00 -6.1837e+00 -8.3177e+00 -2.7277e+00
#>   2.4510e+00 -9.3308e+00 -5.7942e+00 -1.7720e+00  4.6826e-01  1.1992e+01
#>  -2.8549e+00 -1.6654e+00 -1.6712e+00 -7.0679e+00  3.5146e+00  1.0214e+00
#>   4.2725e+00 -1.5950e+00 -7.6010e+00 -6.6313e+00 -7.9974e+00 -9.9291e+00
#>   2.5131e-01 -2.5355e+00  9.5632e+00 -6.3029e-01  1.2606e+01  1.6650e+01
#>   4.6183e+00  5.2454e+00  1.4488e+01 -1.7258e+01 -2.3829e+00 -7.3576e+00
#>  -3.1407e+00 -6.4715e+00 -6.6440e+00  6.0671e+00  1.0910e+01  8.8155e+00
#>   2.2202e+00 -1.6395e+00 -1.0062e+01 -4.2710e+00  1.8262e+00 -2.8603e+00
#>   3.3857e+00  2.6192e+00  1.3864e+01  2.4330e+01  6.5364e+00  6.5230e+00
#>   2.6703e+00 -9.2705e-01 -1.3538e+00  2.7518e+00 -5.5506e+00  8.0936e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]