Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -3.4738e+00 -2.0988e+00  1.3677e+00  1.1301e+01  8.0791e-02 -1.2569e+01
#>   2.5301e+00 -1.0298e+01 -5.5102e+00 -8.9817e+00 -1.2162e+00 -1.3440e-01
#>   3.9376e+00 -5.6538e+00 -6.8986e+00  1.0378e+01  3.6776e+00 -1.1691e+01
#>   1.0565e-01 -8.0097e-02  2.2011e+00  2.0406e+01 -1.3502e+01 -2.5039e+00
#>   1.4674e+00 -7.0376e+00 -7.7331e+00  3.9222e+00 -4.8600e+00 -4.7238e+00
#>  -5.9352e-02 -3.5809e-01  1.3172e+01  1.4327e+01 -5.5279e+00  7.6245e-01
#>   3.8795e+00  3.0490e+00  7.5983e+00 -4.5286e-01 -5.3246e+00  3.0649e+00
#>  -3.4064e+00 -2.2432e+00  5.6880e+00  3.9668e+00 -4.3076e+00 -7.8612e-01
#>   6.2588e-01 -1.0412e+01  1.0695e+00  7.8103e+00  1.3992e+01  7.4433e-01
#>  -2.5550e+00 -5.8952e+00 -9.9833e-01 -5.3885e+00  1.0710e+01 -1.1396e+01
#>   2.2486e+00  4.2708e+00 -9.0870e-01 -1.5766e+01  1.6079e+01  2.9591e+00
#>   9.6815e-01 -4.9776e+00  1.8924e+00  2.4146e+00 -1.8948e+01  3.3001e+00
#>  -9.1044e+00 -4.2383e+00  5.8891e+00  2.5219e+01 -9.9286e+00 -4.4208e+00
#>   1.9628e+00 -2.6451e+00  9.1646e+00 -2.0830e+00 -4.8236e+00 -4.3945e+00
#>  -1.0231e+00 -2.1699e-01  6.9655e+00 -7.3276e+00  7.2352e+00 -1.5513e+00
#>  -8.0751e+00  4.8174e+00  8.2816e-01  1.5467e+01  2.3160e+01 -2.6699e+00
#>  -1.4406e+00 -4.5329e+00 -7.9309e+00  8.9190e+00  1.9252e+01  2.2967e+00
#>  -1.7675e+00  1.4627e+00 -6.2601e+00  3.4837e+00 -1.3878e+01  1.7465e+01
#>  -4.1910e-01  7.7647e+00 -9.6860e+00 -2.5313e+01 -4.9453e+00  1.3464e+01
#>  -4.7546e+00  2.6990e+00  2.6521e+00  1.4249e+01  3.2743e-01  1.4759e+01
#>  -1.3479e+00  4.1393e+00  1.1082e+01 -1.1173e+01  1.0021e+01 -1.0952e+01
#>   6.3249e-01  1.7007e+00 -7.6340e+00 -3.9522e+00  5.2784e+00 -6.4897e+00
#>  -2.7912e+00 -1.3856e+00 -8.4084e-01 -6.7054e+00 -2.0194e+00  8.9396e-01
#>  -3.3617e-01 -1.4042e+00 -3.2736e+00  3.7387e+00  1.9862e+00 -8.6246e-01
#>   5.1022e+00  6.0660e-01 -3.4107e+00  4.9136e+00  6.7952e+00 -1.1320e+01
#>   2.4192e+00 -1.9723e+00  1.7770e+00 -1.0632e+01  1.5194e+01 -1.5741e+01
#>   8.2559e-01 -6.1033e+00 -5.6063e-01  1.3285e+01  1.8902e+01 -5.9405e+00
#>   3.3139e+00  6.6174e+00 -3.9896e+00  1.1147e+01 -1.9731e+01  2.9441e+00
#>  -3.1432e+00 -5.3707e+00 -1.3709e+01 -3.0346e+00 -5.3767e+00  7.9312e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]