Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.4362   7.1062   7.9512   2.8971   5.1031  -5.9420  -6.9622 -12.6344
#>    3.5808  -3.5133  -2.8983   2.6019  -2.7777  -0.8181  15.9716  -3.8749
#>   -6.1201  -5.7548  -5.3206 -10.8777  -0.0641   4.9547   6.1545   3.7973
#>   -3.4718  -6.9697   9.4579   0.5489  -7.6491  13.6538  -0.9545   6.1643
#>   -1.5837   7.5865   1.2323  -2.0000  -2.5297 -11.0921   1.2477   3.4961
#>   -0.7679   1.3052  -3.4540  -0.0464   6.3478  11.6816  -3.2219   4.5139
#>    2.2709   6.2419   8.2958   0.8806  13.9455  -4.2322   4.8966 -10.8811
#>   -2.1923  -4.6578  12.7376   2.2827   2.4860   3.0540  -7.0784   3.8332
#>   -7.3003   1.1748  11.3905   0.1050   1.8867  -6.0186   8.6041   1.5743
#>   -3.8314   9.0107   0.5834  -4.4992  13.8343  -4.6516   3.4392  -9.2077
#>   -1.3255   3.6838  -0.4651   5.6664   2.5521  17.1087  12.0170   5.0501
#>   -1.1072   2.7990   7.3667   8.7220  15.4749  -0.2344  -0.3282  -0.8315
#>   -4.0080  -0.3624   4.4643   4.8866  -1.1408 -13.0829   4.7860  11.6197
#>   -5.4433 -11.0224  -5.2139  -7.7956   2.9260 -10.3323  -5.6717   4.1694
#>    2.0725   3.6764   0.1178   8.5567   6.8062  -0.1876   6.9772  -3.3439
#>   -1.0010  -9.4275  -5.8269  -6.6670  10.1005  -7.3809  -4.5839 -18.8333
#>   -5.1550  11.0990  -5.7312  -9.6299   5.2560  -7.4997   7.6212   4.5992
#>   -0.2369   8.0838  -1.4842   2.9242  -0.6837  -3.6158  -4.0091   8.1657
#>   -3.5059   4.0075  -3.4535 -15.8163   2.9375   2.0730   6.7126  -8.3275
#>    6.4977   2.7600  -6.7878 -17.2401  -0.6759   7.1784 -16.4656 -21.0692
#>    4.9124   8.9130  -1.9684   7.8027   1.0936  -3.6301  -4.4383  -0.0183
#>    0.8328   1.4161   0.4890   5.8470  -9.8247   8.8404   3.4749  -5.5230
#>   -0.7737  -9.8481   0.5896 -12.1970   1.6325  -0.0650  -7.0683   6.8327
#>    5.1265   7.5441   1.9552   0.9403   7.9883  -3.0679 -13.5443 -10.4543
#>   -1.2409 -10.8438  11.0900   4.9168  -1.4554  -4.0857   3.6799  13.3512
#>    2.8879  -2.7603  11.9288   2.9197  17.8678  -2.2276  11.6377  -1.5007
#>   -0.0978   1.8294   6.0522   1.7915  -0.0320  19.6154  -0.6224  -1.1618
#>    1.5386  -8.8934 -12.1679  -1.8177  -8.1953  -5.2441  -1.2456   7.6739
#>   -2.9809  -5.5553   3.6620  -0.5474  -1.4436 -10.4868  -8.8192   3.2888
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]