Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.5967   2.1558  -4.5446  -5.1794  13.8012 -12.6958 -10.9682   1.7351
#>    3.6717   3.7884   7.5686 -16.7787  16.1206   1.2508 -10.4364   8.0030
#>   -9.2573  -1.8776   4.4985   2.7209   1.9185  18.0468  -4.5775   9.5423
#>    1.8246   7.8395  -2.3803 -17.8284   5.1182 -13.7699  -9.1928   5.1760
#>   -2.7151   6.1397   4.2230  -2.7568  15.3493   3.8699  -6.3810  14.2692
#>    3.7512  -6.1325   1.8001 -17.7749  -5.0261  17.8985   9.2331  -0.4509
#>    0.8760  -3.0228   6.2487  -1.6744  -1.1750  -3.8278   5.6580  -3.7879
#>   -1.0811   8.8128   6.8997  -3.1820  14.9999  -5.9194  16.8837  11.3370
#>    0.4335   0.4252  -6.7460  -3.2207  15.3164  14.7164  -5.8826  -6.1067
#>    6.3056  -7.7077  -3.4425   7.5425  -3.3440   5.0569   6.4727   5.8003
#>    1.1159  10.3375   2.6352   5.7565  -4.7398  -2.3940  10.6138  -7.7665
#>   -7.6410 -10.6417  -1.6961   1.0809 -15.4282   3.8788   3.4309 -12.4483
#>    3.3666  -1.1557   1.2944  -1.1668   3.0667 -19.9170   0.8052  -1.4380
#>    5.6110  -2.8095  -5.1563 -16.7069   0.3724   4.9116  -3.7059   5.6570
#>   -5.5610   3.6029  19.5677  -1.2584   9.9465   2.2237   2.5712  -0.7335
#>    1.7330  -5.0462   3.0619   2.7616  -6.0653   4.2552   1.0219   3.7204
#>   -2.9786   4.2024   3.5868  -9.4055  -3.3901  11.4644   7.5144   6.3719
#>   -3.2550   7.3974  -8.4069  -5.4034   1.5947  -9.3094   1.0548  -5.0529
#>   -4.7688  -5.4865  -5.7644  13.2322   7.0377  -0.7859   9.2797  -0.6726
#>    2.0783   2.7149  -6.0045   2.5212   4.4225  -3.6288  -4.3484  18.4253
#>   12.1110  -4.0452  -9.9277  18.1799  -7.9088  -0.6211  11.4749   2.8172
#>    0.1752  -2.4410  -1.1550   9.8071  -0.8575   3.1771   4.6826  12.3077
#>   -8.2684  -5.7166   0.8357   2.8078  20.2501  -1.3564  -1.2561  -4.1641
#>   -1.6036  -5.2465 -12.0953 -10.6543 -18.8970  -4.6096   7.5537  14.4075
#>   -0.7413  -8.6046 -11.3832  -2.5829  -6.2268  -4.1476   3.0574   8.4639
#>    3.2245   4.9332  -0.3782   8.9676   7.1435  11.0522  -4.7491   7.2299
#>   -1.3405   4.1986   9.7217  -6.7401  -7.9181  -8.8925   0.1636  -2.8587
#>    7.0216   7.0094   3.8823 -13.8443  -5.7181  10.5859  29.8461 -21.1966
#>    0.1434   4.5560   1.9684  -2.0862 -11.5494   5.6372  11.0567   0.3465
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]