Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.2240  -7.4029 -10.7628   5.2974   1.0810   0.4662 -14.1521  -0.0303
#>   -0.3501  -3.0292 -11.1795  16.1030  -8.7350  -3.3305   7.8005  14.6540
#>    1.4439 -15.0022   9.8285 -13.7492   1.9301  -4.3413   6.3260   1.1223
#>   -1.6683   0.7218  -4.2786   4.6669   6.8829  -6.6253 -12.9740  -5.8453
#>    5.0182  -0.7904   5.8143   5.6301 -16.8609  -6.7386   3.7466   6.3961
#>   -2.0475  -3.6847  -3.8403   2.3408  -5.7173   8.8275  -3.0915  -1.4196
#>   -3.2740   6.2717   2.4060 -27.0294   8.6773 -32.9913  -5.9889  12.7531
#>   -2.7378  -9.7550   6.8922 -14.8407  15.4291  -9.5841   6.4024  -2.3135
#>    0.1471   2.7767  -6.5117 -10.8679  -8.0903  16.0370  -9.6779  -1.6068
#>   -3.0988   0.1394   3.0336  -4.8849  -5.6570  -6.3915 -11.8684   5.2557
#>   -2.4060   4.4459   1.7961   6.1100 -12.3617  -6.1661  -0.5889 -12.3265
#>    3.4477  -1.5473  -2.1272   2.9030  -2.4588   0.2602  -6.5365  -4.5880
#>    0.2669  -3.8148  17.4206  -9.0669   9.2398  14.6848   3.8297   0.2214
#>   -2.9626   2.0424  13.3743  -2.2669  12.8735   6.3369  11.7254  13.6840
#>   -6.3320  -1.1278   9.2186  -3.4251  -6.3468  10.8859  -3.4942   1.6796
#>    6.7761  -3.6263   6.3931   6.1505   9.6851   0.7546  -2.1458   9.2831
#>   -0.8613   5.4972  -4.0470  10.2983 -14.2195  -4.5506  10.4214   4.6287
#>   -1.0083  -4.8784  20.4751 -11.5975   0.8146  -3.4294  15.8073  -5.1309
#>    3.6413  -8.9379 -11.7588   3.0704   5.4869  -9.9888 -10.3744  -1.7132
#>   -1.2900   0.4395  -9.1061  -7.7029  -3.0076  -2.8522   4.7689   7.4475
#>    3.0967  -6.3346  -9.9420  12.5830 -17.7005   1.2054  -0.1208  23.1723
#>   -2.4354   4.6024   7.7032 -14.0625   5.6425  -9.7544  -0.0997  -8.2983
#>    1.8346  -3.2264   9.5248 -11.9925  -0.6490  11.8076   1.3425  -5.0920
#>   -4.4022   1.0638  -5.8971  -1.6468  -1.9878  -7.5018  18.9049  -8.9913
#>    3.6941 -25.8643  14.1633  -6.8704  25.1192 -24.5740  -6.4222 -17.0078
#>    2.5800   4.3754   3.0919  -4.5691  -4.1573   3.6412 -10.0951   7.8469
#>   -0.7701   2.2209   6.7984   0.2518   1.9595   1.4521   2.9691  11.5344
#>   -3.1764  -2.9120  -8.5043   1.8637  -5.0888  11.4849  -6.8447  -2.8680
#>   -6.7367  -6.1256   4.0013   2.4063   9.6359 -32.0668  13.0361  -0.3002
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]