Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.5891   1.3894  -5.1621   0.7873 -10.0210   7.5392  -8.5898   4.7512
#>    8.6207   3.8900  -6.1961   4.5420   7.2284   8.6449  -7.8910   8.0750
#>    0.5240   2.4711  -2.9250   0.3974   1.1729   5.8072   5.4687   0.4839
#>   -2.8575  -1.3973 -10.5162   1.9766   0.1504 -13.1730   7.9538   0.0399
#>    3.4477  -3.0466  -3.6496   2.4798  12.8445   0.7037  -0.2073  -8.7388
#>   10.0120   2.5471  -4.6679   0.6075   3.3421  -1.4606  -7.8211   3.5947
#>   -2.5630   2.6816   6.7854  -1.4435   3.9509   9.4381   3.1498   1.0245
#>    2.9789  -1.9188   9.0243   8.8838 -11.7116  -3.8649  -5.8441  -2.4320
#>    9.7030   2.6810   2.4822  -1.5798  -3.4022   1.0425  -8.7783   4.3726
#>   -2.0173  -5.3102   3.5667   1.8457   7.7612   2.4782  -3.2606  -7.2124
#>   -4.4435  -0.8108   3.7299  -2.5940  -8.9046   6.7373  -4.4534   6.4432
#>    1.5446  -1.5754  -4.6239   0.5628  -4.2043   3.4439   0.8955   1.9877
#>   -0.1710  -4.2279  -6.9199   5.7190  -1.3027  -9.2178   5.0932 -11.2563
#>    0.2394   1.2957 -10.5730  -6.7538  10.2782   9.4583  -0.4697  11.9944
#>   -0.3696   4.7896 -12.5288   2.2527   9.5234   5.2159   5.8292   0.9867
#>    0.0760  -1.3001  -6.9352  -2.2333   1.7938   0.3351  -9.3671   8.9982
#>    4.2535   5.9360   4.0704  10.5064  -0.1674  -9.7243   7.0758   5.6071
#>   -3.0020  -1.2766   3.0632  -6.6228   5.8118  -1.8899   3.2963   6.4715
#>    4.7384   4.9124   5.7833   1.4272   3.1741   0.3464   0.0872   2.4313
#>   -0.6817   0.8729 -11.0962  10.0252   5.5740   6.0406  -1.4271  10.6122
#>   -1.9447   0.1083   7.2248  -1.3419   2.1390  11.1119  14.6578  -3.4315
#>    1.7153   9.4954   5.5840   4.3147 -13.7421  -4.4474   9.4574  -3.7174
#>    0.3037  -7.8535  -1.2608   2.9785  -5.5110 -10.4829   3.7816  -3.3729
#>    4.0043  -8.4537  -4.5926   9.4007  -0.9853   3.5751  -5.1761 -10.6747
#>    6.0717   2.3567   5.8802   8.9656   9.0310  -6.3838  -6.5012   5.3649
#>    1.9153   1.3388  -5.3598  -9.5570  -4.3450   3.6222  -2.6792  -2.7539
#>    1.1632  -1.8643   7.2502  -7.5978   8.1668  -3.1310   2.5965  -2.2971
#>    1.3867  -5.1004  -2.7682  -0.6029   0.6965  -2.8454  -2.9387   5.4022
#>    1.1395  -0.4795   5.2622   9.6774  14.8805  -7.5392  14.7099  -6.6956
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]