Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -9.6256  -1.9780   2.4761   2.6094  -3.2066   1.4945  11.2962   7.2060
#>    1.5362   8.1812  -4.8297  -5.8945   9.5292  10.3524   2.9964  -7.3201
#>    1.0064  -4.9143  -7.0842   2.4935   6.2916  -4.4263  12.0250  10.4878
#>    0.2473  -7.8360  -9.9231  -6.9223   4.1279  -8.2976  12.6415   7.0674
#>   -2.2887  -0.7646  -7.0097  -7.4371   5.9749  11.6166   5.5757  10.1986
#>   -5.1275   1.8637 -19.7842   2.1190   9.5797   0.6677   2.0830  -3.0864
#>    6.5024   1.9336  -4.6316   8.9704 -13.0957 -12.8162  12.9065 -12.1572
#>    3.5373   6.5241   6.1547  10.4491  -3.4641 -10.8870  -3.6748  -3.7179
#>   -6.5273   2.1952  -2.3448  -0.6174 -11.5039  -2.9361  14.4736  -9.3294
#>   -3.1064  -1.0254   1.7324   6.3434  -2.5071  -9.5954   3.1764   2.4877
#>    1.4601   4.8657  13.6606   8.8125  -0.3957   2.2597   0.6201   9.4009
#>   -8.6230  -8.3833  -7.5775   2.3006  -4.0087   0.4644   0.1744  -3.8780
#>   -8.3651  -0.6584   5.0188   7.8302  -6.0846  -2.2538  -0.4052 -10.1156
#>   -4.4197 -13.7460  -4.8185   9.1077  -6.4705 -17.0744   2.8744   0.5109
#>    3.2193  -3.0552   1.1438  -2.3294  -0.6485 -12.3061   0.7486  10.6753
#>   -1.1730   1.1129  -5.9572 -18.5946 -14.1769   4.5303  -5.9087  -1.2145
#>   -0.3464  17.3601  12.9338   5.9206  -3.2871  -7.3098   3.5636  -4.6480
#>    9.0703  -2.2935  -5.7801  -1.0397   5.7192   4.0970   2.1110   2.7135
#>   -6.7132   9.2539  -4.4256  -7.8446   1.7011   6.1696  12.0684   1.8174
#>    3.3349  -1.3842  -0.5529   4.5774   0.1261   4.1609  -1.1092  -4.0479
#>   -5.3130 -15.0894 -14.1178  -8.2170   0.5130  -7.3098  -3.0484  12.3767
#>   -1.7539  -9.1223   8.0001  -5.0330  -6.3100  -1.1316  -1.3495  -4.7300
#>   -0.5967   2.1044   0.9569   2.5539  21.1382   4.3659  13.3929  -2.1411
#>   -0.9741   1.9281  -3.7955   1.6969   2.3667  -1.4942  11.6348 -13.2944
#>    0.1146  -2.2793  -5.6609  17.6530  -0.3436   0.1731  -4.0059  -2.5622
#>    0.1104  10.2147   5.6804  -9.5158  12.2438  -3.6662  -4.9695  -6.8865
#>   -5.7305  -7.7245  -5.0798 -13.5384  -4.4644  -2.1764  -2.7766   3.9841
#>  -11.2493   3.4812  -6.6915  -9.5620   0.5679 -15.1659 -15.4347  16.8041
#>    9.7292  12.1228  -4.4084  -0.7994  12.1827  -5.9702  10.9392  -0.0581
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]