Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.7932   5.3826 -10.5978 -10.5571  -9.2648  12.2762  -2.3743   4.6241
#>    5.7187   6.6300  12.9787   0.7529   0.8334 -13.6393  -3.1243 -11.1103
#>   -1.7862  -3.1148 -13.1180   3.5442  -7.5778   6.1473  12.3955  -6.3344
#>   -1.0179   1.6651  -2.2829   6.3968  -8.3759   4.5011  13.0896  -8.2702
#>    0.4583  -2.1611  -2.9849  10.3084  -3.1700  27.5525  -3.4001   0.5686
#>   -6.9902  -3.7420   8.9430  -6.1109   7.8360 -10.0506   4.6689  -9.2317
#>   -6.8457  -5.4011   2.4274   1.0840  -8.3095  -5.1630  -5.1049  -4.1559
#>   -4.8042  11.0085   1.1831  -1.6407  16.6437  -5.4197  13.2368   7.0840
#>   -4.9251  -5.7853  10.0028 -11.6847  -2.1326 -10.7790 -12.3493  -8.6296
#>   -5.8782  -1.7090  -4.7961   6.0823 -10.8845  -4.8299  13.2318   1.4367
#>    3.6040  -6.5516   4.7982   0.9083  -2.4104   5.3448  -9.0909   0.2107
#>    5.2348  -7.4192  -0.3290  -2.0066  -2.3886  13.2655  -8.2401  10.9208
#>   -4.1760   5.0416 -11.4277  -5.4370 -10.2774   5.9097   3.7521   2.6337
#>   -0.0415  -0.0687   4.2285  14.2842   5.4561  16.2876 -15.5622  -2.0580
#>   -0.4610  -7.4435   8.5987  -0.9521   2.4035  -1.8891  11.9324   8.5811
#>    3.3373   4.9103  10.2430  -1.0136  -0.4696  -1.4683  -0.3981   4.1954
#>    2.5997   0.3529   7.3525   8.0086  -2.7167  -4.8750  12.5772 -16.2408
#>    2.8410   4.7222   0.0814  15.6430  17.3998  -1.2836  18.0865   5.1422
#>   -0.0992  -6.6484  17.0946   4.4903  -0.6899   3.4611   1.8450  -3.9785
#>  -11.2163   4.2877   6.0474  -3.2036  -6.7708  13.1744  -1.5789  -8.5478
#>   -0.5348  12.3341  -2.2850  -4.8633  13.7982   4.8816  -8.5208  -4.9459
#>   -7.7070   9.2467  -5.3252   6.1297   5.9440  -3.1801   1.9863   9.8162
#>    6.1578   2.8172  -3.4520  -6.3181 -12.2000   9.1022 -10.2928  -9.9293
#>    0.8631   8.5595   0.1523 -12.2087   1.1975   9.0247  -6.5955  18.9048
#>    0.8709 -11.6269   6.4558  -2.8368  -3.8129 -10.8681  -1.4974  -6.2905
#>   -0.9538  -5.5386  -4.1901   3.3955   0.8933  -4.7052   6.3396   7.8325
#>    4.1001   0.9069 -16.5105 -11.0936   1.8310  -5.4881   0.2578   8.9971
#>   -0.4478  -3.0502  -1.7622   1.9629  13.2030   1.8762  -4.2144   0.9307
#>   -0.1790  -2.6894  -2.2222   3.9275  -2.9472  -8.2660  11.6492  -5.4838
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]