Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.1249  -7.3614  -5.2990  -2.0268  -0.9200  -1.6689   1.5797 -10.4130
#>    5.5604   9.1888   5.5021  -6.4515  -9.9761  -9.1955  10.6013  15.0881
#>   -0.4465  -0.1433  -9.4567  -4.8823   4.4910   2.1050   1.3820   4.6554
#>    1.9934  -1.2363   0.2231  -0.6307   6.8274   4.9557  -4.1462  -0.6552
#>   -0.2049  -0.6143  -5.4374   3.1889  -1.4735   2.9161  13.6371  -2.1954
#>   -3.6338 -13.3291  -4.1162  15.1014   7.3293   9.1340 -11.7929   1.8401
#>    2.7800  -1.7828   2.6367   0.4032   5.8450  -0.1767  10.6795  -5.1208
#>   -5.2659   5.1281   5.1256  12.1961   9.4427  -0.2440  18.2930  -6.6227
#>   -0.6005   5.9023   8.1409   7.2603   7.1895  -2.8665  -1.9605   0.3329
#>   -4.3202  -1.7207 -13.6363  -8.4866  -7.7959  -6.3341  13.7731   0.7269
#>    1.7042   5.9729  -8.7347 -11.5626   0.0485  17.4118  -3.9276  -0.2919
#>    0.3159  -1.7554  -8.0385   7.1736  -5.5635  11.0159  -1.6651  -7.7493
#>   -4.5740 -15.6558  -2.3907   5.6260   7.6236  -2.5731   7.8940  -8.4142
#>   -5.9521   6.9981   3.3972   9.7000  -5.8950 -11.3356  -4.3239   4.3205
#>   -1.2810   5.1739   2.7665  -6.2125   6.2630   4.9432  -0.2359  -7.5294
#>   -3.0766  -2.7454   3.7069  -0.3314  -1.0012  -4.5138  -8.0153  17.0686
#>    5.7316   3.3438  -4.6295  -4.8973  16.1528  -4.5465   2.9390  11.5419
#>   -7.1196   4.1475   1.2163  -5.2752  -3.9856   0.5049  -4.4164  -4.5403
#>    6.6692  -8.1016  -1.7712  -1.5629  -2.3405   5.6130 -10.4245   4.0589
#>   -2.8890  -4.0918   1.7745  -9.6885   6.7916  -0.0629   1.8501 -25.9445
#>   -1.4420  -0.2473 -10.0137   9.8556   3.4515  11.5250   7.9118 -14.6198
#>   -4.6666   0.5543  10.5216   2.9587   5.0366  -6.3053  -8.2835  -9.8726
#>    6.0903  -3.0466  -2.6632  -3.1529   3.1391  -3.9608 -14.4080  -6.8956
#>    0.8626  -2.5620  -2.7576  -4.2721  -2.1973   3.9775   2.7514  16.8889
#>   -2.3570  10.7804  21.3286  -0.0156  -1.3636  10.0929 -12.2024   6.3235
#>    7.7323  -4.3235  -2.0557 -11.7488  -5.5302   7.4160  -6.9311  -3.2153
#>   -2.3132  -4.4301  -7.6093   5.8810  10.4639  -0.0648  13.8385   8.2005
#>   -0.6272  -7.5949  -3.0419  -4.2086   8.3772  -0.3337   5.5195   5.3402
#>    2.9322  -4.6917  -3.1787   3.7193  -0.9067  -0.1697  -7.6544   5.7589
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]