Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.3191   8.9100  -3.7966   7.1264 -14.7596   2.8485  -5.0619 -11.8569
#>    5.9059  -2.5986  -9.2496  13.5751  -9.0040 -14.9394   2.9650 -10.4002
#>   -0.2668   2.7124   5.4831  -8.4430   6.4147   6.6234  -7.0627  -7.4149
#>   -6.6104   3.9036   1.7627   1.7205  -0.3526  -3.2928   6.4612  -8.7928
#>    3.7115  -9.3230  -1.9390   0.2439  13.9892   8.2669  -2.4659 -24.6720
#>    5.2829   4.3084   1.9419   1.2765  -9.6053  -0.6191  -3.5789   5.2016
#>   -1.6261   1.5490  -0.4356   1.6466 -14.9190   9.0125  -5.4133  13.2960
#>    6.2377   4.3988  12.2679   0.1856  -1.7030  -3.9345 -22.0130  -5.4158
#>   -3.9484   3.7872  -4.3140  -3.1402  10.0906   4.2271   0.1084  11.2670
#>    1.5488  -9.9225 -10.0941  -7.6364   5.1195   6.8475  -8.8822   6.7156
#>   -3.0136   1.9791  -1.5245   3.8630  -2.5478  18.7050  -9.8961  -7.5677
#>   -2.7140   3.5318  -0.1687  -8.7992   8.0714  -0.9015  -1.4867  -3.5246
#>   -0.6323   2.7124   0.0829  -0.9845  -6.5669   7.4297  -0.0971 -11.4060
#>   -0.0922  -8.6868   5.2655  -5.5474   8.0523   2.9209  -2.9754  -5.4190
#>   -8.4312  13.1731   1.3768 -16.0392   3.9843  -1.0766  -5.1651   4.9903
#>   -6.0781  -5.3758   0.5128   1.0453   1.8920   9.0767  -5.7340  -2.6834
#>   -5.9586   5.2064   4.5808   0.4312  -2.8936   1.1721  10.9736  -7.6473
#>    3.3462   0.8717  -6.0568  -0.8473   0.9523  -9.1925  15.4237  29.3375
#>   -2.3554   4.1149  -3.2708   0.4171  -2.2868  -0.6026   8.0485  -1.7954
#>   -6.0635  -1.3376   9.0442   9.6467  -1.3058   6.4434   3.8774   6.2097
#>    3.2047  -6.9812   1.9550  -4.0789  -2.0851  -1.0084  -5.7009   4.7626
#>    3.2267   6.4457  -7.5146   7.5557   5.3156  -3.9006 -10.0711   7.4996
#>    2.2643  -8.7612  -7.9604  -5.7687   1.8176   2.6684   0.1639  -0.7582
#>   -2.7576  -2.4770   4.8425  -1.4012   2.9507   5.8784  -8.2045  -7.7286
#>   -3.7937  -7.1233  -4.5281  -4.6847  -3.8031   2.8989   1.2457  -1.2372
#>    2.0379  -2.2107  -4.4617  -3.1024 -13.3447  -1.8811   7.1248  12.4926
#>    2.7938   9.5143  -2.2011  22.0988  10.6791   1.5915  -6.6851  -2.6833
#>   13.9066  -5.4436  -5.4452  16.2059   0.7577  13.1593  -0.5858  -1.0755
#>   -1.4629 -10.6439   4.1920   4.9441   3.3848  13.4941  -3.2923  -4.4856
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]