Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -7.2378  -2.8671   9.1276  -9.6621  -7.5292   3.1766  -5.4488 -17.0359
#>    3.6399   5.3203   3.0967   7.0253   4.8504   5.6145  13.6247   3.8100
#>   -2.1453  -8.4974   0.5783   3.4458  -1.4668  -1.0059   9.6470   1.5171
#>    1.7917  -4.0386  -1.1111   4.4329   4.3224  15.7783   8.8816   3.3888
#>   -0.4070   8.5870   6.8251   8.2431  -2.3729  10.0177   0.3695   7.3420
#>    0.5987   1.4423   0.2990   7.6946  -3.5764  -2.1032   1.0139   1.4974
#>   -4.1199   2.7482   4.0054  -3.4000  -6.9245  14.5237  -3.0611  12.7122
#>    3.9302  -0.7444  -2.3623   8.6466   7.0203   6.1585   6.2301  11.5428
#>    4.5665  -2.6561  -0.9197  13.1995  -3.4891   8.2018  -4.3832  17.1133
#>   -4.2167   4.9747  -2.6913  -1.5624   6.1250   8.7438  12.0047   0.5004
#>   -0.1066  -6.1314   2.3245  -6.5544  -3.4664  10.2415  -3.3095  -4.1456
#>    7.2605 -12.0139 -20.0915  -2.1036  -9.1027 -10.9199   0.7801  -4.0171
#>   -2.3985   3.8653   5.8823   1.9874 -15.4008   9.2659 -13.7834  -2.9819
#>   -2.3431  -1.7723   7.6580 -11.0626   7.9377   3.2333  -3.9914  -3.4611
#>   -7.1412   3.9284   5.2028   1.3344   3.8274  12.0368 -10.8683 -13.6592
#>    4.8759   2.4941  -8.0262  -2.4997  -0.1695   6.2238  -1.4234   5.0421
#>   -1.5344   0.8187  -3.2293  13.2822  15.0362   3.4444   3.7559   9.8311
#>    2.0345   2.2564  -2.7902  13.2076  -7.3957  -2.3239   8.6033   1.1880
#>   -6.5855   3.3549   5.7535   2.6433   3.4973  11.4258  -9.1615   7.5659
#>   -6.5908   9.6027   6.8994 -11.7770  -0.5165  -2.1409  11.3801   1.2784
#>   -3.4095   1.9547   6.7948  -5.1693  -1.4168   8.0087  15.4204  12.6507
#>    1.2077  -9.3082   3.7325   8.8713  -5.1551  -7.3435  -0.8185  -7.2464
#>    4.2880  -0.5634   8.6143  -7.2525  -3.3391  -9.4970   4.1078  -5.9889
#>   -1.0725  -1.2477  -4.5136  -2.3785  12.4493   9.8559   6.8332   5.8497
#>   -3.6364 -10.3255   2.1541  13.4103   6.6239   9.2896  10.0763   4.0073
#>   -1.8192   7.4912  -0.1239   4.6603  -6.8187   8.1909   7.0496  -0.8512
#>   -5.5019 -12.4503   5.1722   3.8436   0.4957  -0.4961  -2.1170  -3.0738
#>   -2.4996   5.5854  -0.4339 -19.0671  -1.2059  -6.3930   0.0958  17.3177
#>   -0.0940   2.1855  -4.7130   1.6406   3.6512  -0.6227   4.7539   3.9983
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]