Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.7307  -1.3298  11.5489   5.3677  11.6640  -0.5270  12.1075  -7.5920
#>    5.4490   6.7116  -6.5711 -11.9682   2.3496   7.7939  13.5785 -16.3873
#>    7.8963   1.3047   7.7647  13.2098   3.6443  -3.6643 -16.1378  11.3652
#>    1.8204   3.4442  13.6718  -4.7822  13.9734 -11.3756  10.5911  -7.7805
#>    1.8816  -3.5105  -7.5861 -16.9869  -4.0889  -3.2693  13.0320  -2.4755
#>    0.3297  -5.3984   1.3676   8.2389  -1.3510   5.3240 -13.2636  -0.4315
#>   -3.2176   1.8083  -7.8972  -5.6384  -1.1167   6.1911  10.7779 -22.2143
#>    4.0324  -1.0766  -3.6423   4.4073   7.4285   6.9582   0.2609  -4.2998
#>   -0.4581   2.6586   6.1999  11.2221  -9.3582   9.7512  -9.6133   3.9836
#>   -1.1522   1.2811  -2.1392   7.0114  -8.3406  -4.7364  -3.1674   3.9049
#>    5.1090  11.7392  13.4982   3.5538   1.6526 -14.4177  -5.9950  -4.2825
#>   -1.5582  -4.9774   2.1780  -9.4762   9.8905  -8.9357  -1.1809  -2.6710
#>   -5.5380  -6.1565  -4.0656   1.1759   5.6149   4.2497   1.0167   4.7888
#>   -0.3542  -0.5531   0.1953  -3.1214  -4.6244   4.9547   1.5574  -0.0524
#>   -1.4401   6.4405  -1.7834   2.1087   8.1486  -3.7702   2.6437 -17.3539
#>    2.1802   0.0074  -3.5887  -0.2172   4.8701 -14.1303 -16.2463   1.7924
#>    3.0786   2.1841   5.0946   6.6150   3.2663  -6.2823 -14.1965  -6.0037
#>   -5.1606   2.2832  -4.7537   6.6884  -6.2687   1.3943  -7.4386  -6.4973
#>   -3.7777  -2.8509 -13.9455 -12.4442  11.7914  -1.1626  12.1071   2.7016
#>    1.8933  -2.0980   2.9041 -13.1303   8.4235   2.3259   0.5206   9.5054
#>   -3.8952  -0.3543   2.8107  13.3863  -8.1913  -4.5742   2.2407 -13.0044
#>   -6.2732  -4.9370  -6.9557   4.2374   5.8442  -3.7531   9.1888 -15.9470
#>   -2.1464   1.6617  -2.4036  -1.2740  -8.8960  -1.0885   0.6985 -12.4628
#>   -0.1554   0.9029  -1.4228   1.9534 -10.5989  -1.9205  -1.7801   6.4321
#>    1.3900   1.2005  -5.5439   6.0119 -11.8184   7.7411 -14.0047  -1.4789
#>    4.4276   2.0656  -6.5246 -22.6776  -3.7985  -5.1793  -4.9499  15.4332
#>   -0.6648  -5.8715 -11.5022  -9.1463  -1.5814   4.7559  14.3840   1.4104
#>    3.3545   9.7468  -5.3714   1.0908 -11.3640  10.0157  -0.7848 -10.9784
#>    0.4400   0.1463  -8.8115  -9.6991   5.1049   9.3773   9.8217   5.4502
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]