Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.1039   8.0089   4.3827  -4.0571  -5.3924  -1.1098  -4.1621   0.4561
#>   -5.1167   0.6322  -5.1757   0.0136  12.6934  -8.8765   6.3224  -7.9034
#>   -2.4243 -10.2990  -0.0362   9.8050   8.7311  -3.2203  -5.9631   2.5488
#>   -0.8786   1.3809  -9.6125 -10.5054   1.2915  14.5833   0.7781  28.5082
#>    3.9437   5.6474  -3.5786  15.4153 -11.9626   0.0195   0.5009   7.3573
#>   -4.1589   2.1440 -13.9704  -3.8368   2.5770   2.6788   5.6428   2.0969
#>    2.3931   2.4054  -0.1039  -5.1563   0.8075   7.5396 -25.7834  17.7059
#>   -2.0043   0.2431  -0.0139  12.4322  -2.2704 -10.7462  -7.0502   0.2127
#>   -3.7090   5.2118   2.9181  -5.5284  -8.6636  -8.1679  25.8934 -10.1632
#>    6.0887  -0.0312   3.4776   2.0701   4.6643 -10.3806  -7.9178   6.1934
#>   -3.8627   4.2413   3.5973   1.1530  -4.2649  10.1152   5.1352   2.3558
#>   -0.8030  -1.1050   1.4918  -8.3844   2.4044 -16.6926   9.6708  -4.9825
#>   -4.3646  -3.1130 -11.1352  -3.5672   3.3397   7.0021   2.8654   2.9836
#>   -1.9521   2.9254  -5.3904  -1.6236   6.4023  11.6692  11.1578 -17.8525
#>   -2.2829  10.3978  -0.1513  -2.6586   1.3820  -0.0856 -17.1535  -7.0676
#>    4.2962  -1.1508  -0.2389   4.5507  -3.3227  -0.3322  -1.8417  -2.0410
#>    1.1866  -2.1766  -5.9932   1.4419   9.7286   7.3468  -9.3423  10.9893
#>   -2.1691   3.9860  -2.9901  -1.2525  -4.8244   8.4350  -1.1082 -10.8044
#>    1.7555   0.7790   9.9341  -5.6303  -6.0472   6.4616  16.5992  16.4094
#>   -0.7892   0.1171  -8.5178   3.7884  -3.0544  -7.8582 -11.4572  -9.4728
#>   -1.3125  -3.7428  -5.7718   4.1214  16.0993  15.4469  -1.9746  -6.3928
#>   -0.0073   9.9949   3.6935  -4.1009 -18.3868  -8.5030   0.9901   7.2264
#>   -0.8419   5.9345   1.8556   9.5242  -9.0645 -10.9194  -3.2730   1.4360
#>    9.1140   8.3929 -11.0705 -11.1568  -8.1687   9.8977   1.2820   3.1709
#>    2.8231   4.8824   0.5367   1.3591   0.4020   8.2350   4.4235  -2.7055
#>   -1.3422  -2.3954  -4.9067  -8.7048   6.8430  14.5837   5.8683  -8.5734
#>    2.0703   7.2592  -1.2164 -10.6976 -10.5435  -7.0596   9.0000   7.7043
#>   -0.0844   1.4979   0.3533  16.6581   6.7424   5.1974   1.3600   3.0675
#>   -6.1075  -0.1528   9.8235   4.4993   9.2614 -15.0750  -3.1222   2.6518
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]