Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.1559   4.1033   9.0991  -0.7071   0.9949  15.1651  -7.9280   1.5310
#>   -2.5833  -8.4513  -4.8989  12.2117   8.3059 -14.6389 -11.9017  -6.9535
#>    6.1323  -6.5562   3.1963   2.5755  20.0215  -0.2540  10.0725  -0.4577
#>   -2.7059   5.1775   4.1851 -10.0524   8.4567  16.6251 -15.1071 -14.2709
#>   -0.1275 -12.5180   0.8533   4.2941 -10.8126  -1.1988  -0.4046 -14.5851
#>    4.0136  -7.7153   4.3705  13.0656  -1.5011  -1.5748   7.8083  21.9377
#>    5.8163  -1.4221  -1.0898  -0.4952  -1.4213  -1.6707   1.0576   9.6837
#>   10.1896   1.8706  -4.9159  13.8261   3.6350 -14.8472   8.4925   9.3435
#>   -5.0970   3.8634   1.2497  13.2824 -14.8056  22.0046   5.7170 -25.3916
#>    2.0832   4.7957   7.2916 -11.4556   7.9249  -3.8519  -4.1476   3.5410
#>    3.7524  -5.3566  -1.7214  -4.8008   4.0539   3.4815  -8.6606  -3.3080
#>    8.2795   2.9078  -4.0228  -7.7861  11.3090   0.7920  -5.2028  17.5040
#>    0.5013   0.0472   2.8324   4.3275   3.8755  -3.3056  12.5721   5.6547
#>   -0.8499  -0.0845  -4.8868   1.5658  -5.2533   6.2602 -10.7209 -12.2653
#>    3.7355   0.2059   2.4858   7.7681  -2.9336  -6.7114  17.8138  -9.8983
#>    3.4437  -0.5586   2.8156  -4.4699  -2.8479  -9.3396   6.9182   7.6545
#>   -2.7646   0.7822  -2.4119   3.8716 -10.8720   8.0646   5.8241  -6.4958
#>   -2.8705  -3.2786   4.3877   7.3064   1.6735   3.1009 -10.4757   7.9773
#>    1.5063   2.0342   0.7713  17.6962  -9.4786  -3.5787  14.3250   6.4885
#>   -6.3237   2.5699  -0.8154  -6.1285  18.4464  10.3325   2.2615   0.0890
#>    0.1425  -1.2609   4.9011  -1.4400  -5.2587   3.4816  -3.6823 -19.3889
#>   -0.2618   6.5522  -5.7080  -4.7321  10.6180   8.8053 -13.4718   6.4194
#>    6.1789  -0.4341   8.1483 -12.1012   9.1569   7.0453   1.5490 -13.1219
#>   -7.6570  -4.4902   4.2726  -1.3431 -12.1272  -3.1394   7.7487 -10.8684
#>   -1.7663  -4.8816   7.0203  -1.1800   6.9027   8.2829   6.7285  15.4120
#>   -2.4428   3.0599  -4.2500  -3.8133   1.4301 -25.2111  -7.6923   3.2082
#>   -2.5626 -11.0176  14.9602   1.8665 -13.7519   8.5861  -7.0449   1.1699
#>   -2.1741  10.5931   9.1786  -7.5072   0.6065   0.2880   4.8500   4.2917
#>    1.0409  -4.3017  -9.2464  -6.5322  22.3477 -20.1786  -7.1886  11.7367
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]