Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.1410   5.9172   0.5635   7.6804  -3.9412 -17.4832  -8.6085   3.2158
#>    8.8099   2.4654   6.1408 -13.7773  -0.8107  -4.4133   2.0175  22.3973
#>   -0.1804  -2.4145  -7.8935  -4.9350   6.3106   0.8529  10.1167  -5.5320
#>    9.5767   5.9022   0.9336  -1.3535   6.4089   0.6105   8.2221   6.3332
#>    1.1788   0.6883  -8.1872  -0.9972   2.8635   2.5197 -10.3445  -5.2013
#>   -1.2339  -3.5934   0.7036 -13.1011  12.9600 -11.6947   6.4698   9.3031
#>    0.9112   3.3228   3.8714  -7.0533  -7.6371  -8.8075  -4.8726  -4.8935
#>    2.3587   9.8277   9.2019 -13.7068   0.8410  -6.6581  19.5028 -18.3382
#>    2.2818   4.2800  -5.8341  12.0483   2.7145  -1.9166   6.7084   9.5286
#>    0.5441   8.4590   3.8107  -2.8202  21.0409   2.7203  15.0026  -0.9883
#>    1.1422   7.8946   2.4073   6.8305  -0.0328   7.8318  -2.2454  -3.4559
#>   -3.3759   3.7042   7.1176  -4.4675  30.5870  -1.0163  16.6092 -17.5820
#>   -4.8739   3.4261 -13.1422   2.3069  -3.1548   2.0085  -7.8965   5.2225
#>    8.9723  -9.5794  16.0685  -3.8620  -6.2015   4.5095   8.7824  -8.1949
#>   -1.6176  -1.6087   6.3506 -10.0201  -4.0767  10.0218  -7.6016  -1.8565
#>    1.8906   3.1304   9.7207  -2.3822   2.6620  -0.8782   9.8835   8.8643
#>   -1.0621   6.2210   3.0553  -0.0637 -10.5922  -0.0780  -9.1972   8.5061
#>   -4.0494  -1.7950   6.7310   0.8901   5.7997  13.6885   1.1381  20.2381
#>   -5.1129  -7.0731 -13.1454  -2.6839   2.3042  11.4691  28.7023  13.9759
#>    0.2289  -7.3891  -4.8461  -4.8975 -17.2334  -6.8611  -9.5265   5.4687
#>    0.7443  -0.9362   0.5446   6.4891 -12.6346   2.8769 -14.5611  10.7288
#>   -0.4506  -2.2915  -3.1273  -1.1345   6.5297  13.0804   1.9657 -10.5475
#>    3.3380   2.0400   7.4036  -0.4379  -4.6235   2.4915  -0.9306  12.4173
#>   -3.0400   3.8138  -9.1012 -14.6056  -6.5406 -11.2184  -3.2715  -2.1333
#>    6.7927  -6.0761  14.2031  -3.8749   0.9070  -4.8511  -7.9525   3.1642
#>    7.4038   1.4026   4.6957   0.6418   2.7958  11.8924 -12.3129   8.0360
#>   -8.4352  -1.5042   2.2712   8.5806  15.3858 -10.0726   4.6207  -3.9798
#>    0.1844  -1.3855  -2.9144   2.7498  -3.9657   5.1158   3.4682  -3.0631
#>    6.9786  13.1731 -11.4340   5.4935  -7.2053  14.1563 -11.5171   5.6401
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]