Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.5954   5.6642  -3.2363 -11.7767   9.1522  -1.6191  -0.4789 -17.5285
#>    3.9939   0.2153   9.4306   2.0506  -7.4209   7.4960   7.5234  -4.2409
#>    3.3233   8.3918   4.4376  -2.2440   6.3409  10.8108 -16.5516  10.1244
#>   -2.2842   8.8450  -2.4566   0.7198 -11.2500   2.1742   2.8438  -7.4099
#>    1.2830  -5.4556  15.0203 -15.4825 -15.7740  -9.3856 -10.9520  -3.7394
#>    4.8794  -1.8304   0.0177  -8.2077   4.2430 -14.9609   3.9702  -8.2796
#>   -5.3319   0.4381  -6.6013  -5.5195  -5.9918  -1.9549  -0.6563  14.1301
#>   -2.2444  -5.8893 -10.5584  -3.4506  18.0251   1.7251  22.4177   9.3193
#>   -0.3771  -7.3020  13.3165   7.4838   3.0774   5.4325  10.3383 -25.4208
#>   -1.1556   1.7049 -11.0940   9.8268 -10.8699 -15.0918  -1.6809  -1.9684
#>    1.4933  -0.1704  -4.6025  -3.7625 -11.7443   1.8086   5.7972  19.4034
#>    2.0212   2.4914  -8.1094  -4.7390  -7.4715  -6.8361  -3.4815  14.0016
#>   -4.4470   4.5027 -14.6478 -12.2056  -9.5749   8.2081  -5.8994 -13.2223
#>    0.7550  -4.2394 -11.0614   0.3885 -20.7711  -1.0659   1.8896   1.4450
#>    0.9730 -11.3156  -4.9489  -6.7946 -13.8675   2.4392  10.1268   5.3454
#>   -7.5675   8.8036   0.8168  -5.2248   8.1181   0.4840  -7.5395   1.3341
#>    5.0978  -2.4895  11.4952  -7.5147  -0.5859   2.3290   2.7665  12.9658
#>   -7.9170   2.1757   2.0809   5.0800  -1.5664   9.0692 -20.9931 -17.7503
#>    4.4919  -8.8375  -4.7124  -1.7142 -10.5032   1.0554  -4.1403 -13.1740
#>   -4.8275  -4.2794   2.7867 -17.3805   2.4486  -4.7504  -0.0831  -3.9808
#>   -4.2479  -0.4329  -2.2735   0.1989   5.6729  -0.1895   6.8427  -7.5496
#>    1.9541   4.0390   8.1131 -13.1926   5.7861   9.1017  -1.5000   5.3013
#>   -1.5672  -6.1777  -7.2534   3.8288   4.8173   2.7464   0.2826   4.3492
#>   -2.0662  15.6048  -3.7827  12.0925  -8.7248  -0.1062   4.6347  -2.0464
#>    0.5698  11.6494   9.1318 -20.8653  -8.7975  -8.1036  -5.2930   3.8754
#>   -7.2939   3.6282  -5.0577   3.0903   3.6833  -6.1383   2.9825  -8.5774
#>    5.6241   6.0448   8.6191   9.9822  -3.1407  -1.5021 -16.5523   1.4351
#>   -2.3181  -2.6540  -0.2080   5.5485   3.7691   5.4941   2.3599  -3.4153
#>    2.4568   1.9910  -0.9856   4.5583   0.4246  -2.3002  -2.1127 -13.5017
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]