Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.1216  -9.8342  -5.2869 -10.1359  -5.1703  10.7258   3.8042 -22.7091
#>   -6.7275  -7.1188   8.8772  -5.7494  11.7983  12.6326  -9.1364   1.9651
#>   -1.5163  -6.5084   2.5562   1.0715  -0.2451  -2.2611  -5.1112  -9.5738
#>   -5.3927   2.8373   3.6060   7.6548   7.8623   0.7007  -2.5901  -0.4158
#>   -0.1712  -5.0176  -4.9449  -9.8142 -10.7430  -0.7010   3.0186   2.0536
#>   -1.3470   1.6775 -10.0018   4.0922  12.9002   0.3376   2.3388  -5.5360
#>   -0.4430  -5.0397  -4.5008   2.3238   3.9284   4.9680   0.4079   5.0947
#>    4.6753  -3.6304   3.8420  -2.0798   1.2570   1.0310   8.0013  -5.7275
#>   -3.4631   7.0658  -2.1867   5.6411  -0.5132  -4.4752 -10.5712  16.3588
#>    8.4602   3.7679  10.5953   6.8002  11.7587   5.5643   0.1715   7.8172
#>   -4.7375   4.9229   0.1948  -3.2947  -9.7337   2.8720  11.8922   5.3383
#>    2.2521   1.4315   0.7136 -11.6186  -2.1095  -2.4923  10.3569  -8.4747
#>   -0.6020  -6.5022 -12.7277 -12.4177   5.3987  -0.9482  -9.9579   1.5562
#>    4.0949   5.2700   9.0096   7.0605   3.5890   3.9692  -1.1116  -5.6360
#>   -4.1524  -5.1024   7.6328  -0.1534   1.1785   7.5856  -2.6090  12.8543
#>    2.9736  -6.3572  -8.2530   2.3834  -7.5318  11.1932  -5.7634   3.6983
#>    4.4035   9.7918  -2.7647   7.1835   1.2395   2.3148   4.1769   3.2011
#>    4.5679   1.1932   7.9527   0.1427   8.2099   6.3266  -4.9269  -7.2300
#>    7.3914   5.7907  -0.0919  14.2237   7.3791  21.4652 -23.2106   8.1467
#>  -11.6599  -4.8104   1.3377 -11.2664   2.0984  -9.1123 -15.4911 -17.5087
#>   -0.8258   0.5313  -9.2552   0.5503  -9.7953  13.0719  -4.2158   0.8190
#>    2.6936  -9.8267  -6.8188  -3.7747 -12.0447  -3.3447   1.2455   1.8492
#>   -1.4794  -2.5474  -1.2268  -7.1510 -15.2830 -10.6264   1.4925   8.3014
#>   -8.6454  -8.6636  -5.7599  -8.8844  -8.2081   3.2857  -9.5073 -11.0324
#>    0.5623   0.9394   1.4449  -4.6927  26.3440  -0.8725   8.6535  -1.1923
#>    1.1028  -6.0659  10.1556  -2.2954  -1.4834   4.4551   6.2217   8.4818
#>   -1.1501  -2.7899   1.3766  25.6742  -2.1630  -6.9036  16.8954  21.6174
#>   -7.3142   0.3285   3.2614 -16.9869  -4.9022 -15.5607   1.9569  -9.4019
#>   -4.7412   0.2361   2.0244 -12.1632  -2.1827  -4.9559 -17.2933  -9.5676
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]