Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.4419  -8.3689  -5.6145  -9.0005  -2.3741  -0.0370  21.0026  -6.1163
#>    0.9030  -0.2091 -12.8239 -11.8128   7.7040  -4.0648  13.2321   0.4429
#>    6.1706  -3.9269 -13.8757  -4.7057   0.8020  18.2601  -2.3554  -9.7924
#>   -1.7095  -5.9684   1.5097   1.5338  -1.3244   1.7350  -6.4227  -2.5096
#>   11.3464  -3.8541  -8.6318  -7.3164  10.9417 -14.3056   1.5454   4.7151
#>   -2.0069  -5.6110   1.6078  11.1617   4.6697   8.2509  -4.6473 -23.5264
#>   -8.1111   0.3869  -4.6411  -5.1014   3.2166   8.2888   5.0929  17.4736
#>   -1.3355  -4.3580  -2.2523 -11.0174   4.7762   8.7962   1.9509   9.1807
#>    3.4017   0.9269   1.9623   9.9906 -16.5627   3.4873  -2.1494 -11.5146
#>    6.9166 -10.3205   4.2063   5.5947   1.2481   4.0450  -2.4551  -4.7167
#>   -1.8920   6.9434  -2.3174  -3.9514 -16.4906  -1.5039  -0.2879   6.2085
#>   -5.1248   3.7458  -6.1273 -12.4457  -0.1105  -8.9590  -7.1027   5.3026
#>   -3.3465  -3.9452  -4.4727 -10.5175   6.6216   0.9139   0.4168   1.8601
#>   -6.3515  -2.5548   6.3767  -5.5198  22.6814   7.7483   4.4190  -3.3458
#>   -2.1598  -0.4990  -0.8122   5.7896 -15.5959   7.8205   3.8163  -5.8759
#>   -0.0530   5.1327   4.5793   6.9358   0.4967   4.4354   0.8098   6.6287
#>    2.4177   0.5521 -15.4296  -4.7164 -16.3638   2.2964  -9.5116   3.9471
#>   -6.6506  -0.4651   2.8268   5.1576  -8.7503   7.1645  -3.1079   0.6355
#>   -0.5051   7.7651   0.0504  12.2933  -5.9036  -1.2407   8.5457   1.7782
#>   -4.1309  -4.8807  -2.3723  -1.7202  14.3396  -2.4585  -0.6191   2.8107
#>    3.3188   4.0531   4.3663  -8.1037   7.3976 -10.1650  -2.4211   2.7735
#>    2.3026   2.9005   5.7846 -15.8633  -8.5942  -9.0099  -3.7491   4.2903
#>    4.1587  -2.5731  -6.2785   4.0613  13.3154 -11.9607   2.1207  12.3010
#>    4.9828   3.8843  -6.7024  -6.5609  20.5182  -8.5226  -4.9585  -5.5090
#>    5.4261  -9.9344  -3.5290  -5.5058   9.8516  -3.3010  -2.8535 -11.0605
#>    3.4520  -0.6526  -1.0077   3.1212   3.2754  -1.3051  -3.2633   3.5321
#>    4.2519  -0.4131   1.0902  10.5056  -4.1456  10.7089   3.0587  12.1317
#>    2.1555   2.7871   3.7924   8.0409   2.2637  -7.9631  -9.0649   1.6252
#>    0.0488  -6.2569 -12.2894   1.9546  -4.1633  -2.3412   4.2395   0.7343
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]