Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.0763  -4.9314   2.6981   0.6490  -7.5819 -10.6757  -9.3890  -2.9869
#>    6.6002  -2.5105   5.4993  17.4542  -6.9863  -5.1024  14.1800  -4.0640
#>   -2.5377  -0.1490  -2.2518  16.9831  -8.6203  -6.5840  -0.0628   0.6915
#>   -5.5073   0.7752  11.2723  13.3843  -2.4448  -6.0634  -9.2183  -9.2746
#>    6.7490  -4.4689  -0.5083  -5.5457  15.7901 -15.4660  15.6572  -6.0485
#>    3.1440  -1.3859  -5.3351   2.5561   0.9706  18.1352 -17.2389  -3.8074
#>   -3.6992   4.3235   1.4364  11.3457   1.7616  -8.0254  -8.4284  -8.1758
#>   -1.5088   3.3160  -3.7488   2.7697  -4.8049   3.2135  -1.3328  -4.5687
#>   -3.1415  14.8314   6.7835  10.7103  -5.0449 -11.5510  -1.8681   6.3482
#>   -2.4585   7.8135  -9.5978  11.9479  -6.7079  -8.6513 -11.4766  -6.3822
#>    3.0358  -8.8819   9.7366   2.5429  -4.2461 -11.3917   2.0103  -7.3878
#>    3.8165  -2.9160 -14.3658   9.8150 -11.4874   3.7646  15.1746  -2.7540
#>    3.2562   3.2186   1.4953   5.5198  11.2984 -12.4504   9.3530   2.4349
#>   -2.3756  -2.0562  -4.3465   1.6839   3.8398  -9.6307  -1.2840 -10.5952
#>   -0.0194  -3.3387  -5.9071   8.9762  14.3467  -4.3826   7.5270   2.0402
#>    2.0764   5.1875   6.0925   0.4393   9.0053  -8.1361 -11.5487  -8.8235
#>   -1.6629 -10.5556  -9.4039  -4.6136   1.9388   3.2875  -8.7517  -5.2122
#>    3.0736  -4.3095  11.3682  -5.6852   2.5002  -9.0400  10.6326   4.0446
#>    2.8608  14.0271  -0.7991  -5.7178   1.9697  -3.1626   6.3659   0.2673
#>   -0.3388   1.5476   2.1525 -12.6297   3.0258   0.4605  -7.7978   2.0639
#>   -2.0238  -0.5357   2.1501  -1.8483  11.1592   2.6669  18.5217  -8.3715
#>   -4.9792   8.0222  16.4428  -6.8982   2.8036 -13.1365  -3.0695  -5.6357
#>   -0.4163  10.9620   0.2987   7.6284  -7.6957   6.3753  -0.7313   6.2182
#>    3.4743  -2.9183   1.2303   2.5251  -2.1912  -8.6794  -5.1545   0.8307
#>   -6.5533  -6.2873   5.5199  -8.5860  -1.3327  -5.9544  -4.7498  -4.3629
#>    0.8720   0.1865 -12.1689   3.8976  -5.8092   0.7052  -9.1539   0.9288
#>    1.5366  -4.4603   1.5315   4.9222   2.3282   5.7139   5.5608  12.5867
#>   -1.6252   2.6443  -5.9190   5.1042  -8.2372  18.8334   5.6926  -6.7627
#>   -4.5265   3.6198   1.2440  12.8262 -25.8522  13.6611  -0.2598  -1.2211
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]