Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.2969  -2.8545  -0.5567  12.1628  11.2971  -3.5053   7.4395  -2.3051
#>   -1.1247   0.0019  -7.4262   5.1583  13.2003  15.5742  -1.0485  -1.0419
#>   -1.9816   2.3011   8.5242  -0.2788   1.6386   5.6797   5.5101  -0.1849
#>    2.1786  11.2238  -6.1342  -5.9216 -16.9265   5.8958   3.7731   9.3068
#>    4.9994   3.5139   3.1977   6.3126 -10.0090   1.4038  -6.7037  13.9816
#>    1.5869  -3.3056 -11.3969   7.3391 -18.9022  -6.0273 -14.0553 -17.2566
#>    2.0733  -0.8893   4.8925   4.2184  -8.3953   4.5675  16.5249   1.2213
#>   -2.3770   1.2234  -6.2964  -7.3102  -0.6308  -1.7128   3.7671   5.5632
#>   -5.4888  -1.5505   0.5992  -8.3186 -12.7534  -0.3394  -4.0691 -16.6695
#>    2.9715   1.1049  -3.3434  11.2655   3.0975 -20.9435   7.8185 -18.7987
#>   -1.0425  -4.4496  -2.1054  -0.5318  11.6794  11.8343  14.2215  -2.6212
#>   -0.7173  -0.2826  -2.4401  -0.3572  10.7588   9.0324   5.1019 -14.8844
#>   -6.5453   0.6889   0.2837   6.2814   4.7081   3.5017   2.0419   3.5759
#>   -2.7999  -5.5441   6.6510   1.8841 -24.6005 -12.6143   6.5818 -13.6268
#>   -5.9746  -3.3130  -7.2285  -1.1402   8.4143  -7.5228   1.0105 -10.9764
#>   -4.1766   4.3757  -2.5217  -7.0350  10.4401   9.0428   7.5073  -2.9393
#>    6.4232   3.7744   5.6794   7.8066  15.0312   9.2661 -23.0082  10.0895
#>   -1.0600  -1.9006  -6.3034  -5.2247  -1.5013   3.7498 -13.3480   3.7325
#>    7.2723  -6.4204  10.6086  -0.4686   2.3146  13.4538   2.7232  -6.9925
#>   -3.8490   9.5639  -3.3744   0.0580   0.7339   3.5101   0.0182   5.8069
#>   -0.6041   6.0896   8.4844   3.0164  -7.0850  -1.1939  -8.7364   0.7511
#>    4.3505  -0.6255  -1.2188   1.2867  -0.2754  10.5828  -0.9504  -2.5125
#>    3.3576  -3.5263  10.3408  -2.6994  11.0909  11.3337   3.4715 -10.0131
#>    2.4993   7.0156   1.8886   3.9568  14.4846 -14.8575   9.6057  22.4153
#>   -1.4842   1.8488  -1.3965  -6.1071  -0.3935  -3.2903  17.2544  13.7721
#>   -0.0224  -3.6230  -3.4360   2.3841  -0.8713  -1.6595   0.3339  -5.6722
#>   -6.5079   0.9689   4.3380   3.8617  -2.7217  21.6379 -11.9361  19.9918
#>    2.4645   3.4389  10.3662  -5.5724  -4.9753   3.2870  12.0871  14.0112
#>    0.0551   0.4351   5.7227 -10.7548  -9.9430   2.0428  20.1461  -1.9232
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]