Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.7340   0.5185 -15.1338  -4.9140   6.5782  12.9965 -12.4973  -6.9653
#>   -4.6941   0.1252   4.4183  -6.4190  -6.5184  -0.8696   4.7965   7.9831
#>   -1.5704   0.2842   4.3864 -12.2306  -3.3995  -9.4454  -5.4758   5.7601
#>    3.1780  -1.0603   4.0246   6.2502 -10.8244   5.2037  -1.4682  -0.3343
#>   -1.0652  -1.6312  -7.7861  -9.5818  13.7578  -2.6077  11.2105  -7.4263
#>   -0.7121   0.3571  -7.0911   1.7168   4.8676   6.9426   7.5697  -7.1284
#>    1.5294  -4.4158   5.6665  12.0089 -13.5750  -5.6684  -9.7094   5.1659
#>   -1.7942   1.2041  -7.5241  14.2983   1.9837  -9.9177  12.6491   0.5495
#>   -4.8262   4.1941  -0.6339   1.0735  -0.2761   0.1824  -2.7120 -16.2547
#>   -6.2994   0.3125   3.2174   6.3811  -4.3125   3.5723   9.7888  -2.6119
#>    0.8190   1.5738  -4.8669  -7.0522   4.3564  -7.6648  13.7426  12.9462
#>   -5.3429   2.7742   8.2105 -13.1147  -6.0977  -7.2330   2.2522  -0.2713
#>    2.5679  11.4866   3.9110  -2.6291   8.1900 -12.6952   8.8871  -2.1185
#>   -6.6145   4.5759  -4.3892   4.3514   8.8060   5.1752  -1.3765  -0.1597
#>    2.8871  -1.6302  -1.8314  -2.5323  -1.5939  -5.8026 -15.1140  14.8745
#>    1.2491   5.9177   4.4948   1.7544  -6.6386  -3.0223 -21.1728   3.5129
#>    1.6415   4.4711  -2.8377  -9.0336   1.8224  -1.4515 -16.4444  -0.8211
#>    6.9361  -1.2170 -11.6026  -2.0550  12.6723 -13.2871  -4.0504   9.9761
#>   -6.3979   2.1881  10.6170   3.9425   0.1455   8.6313   2.4141  -9.5627
#>    3.6079   4.3788   5.4856  -3.8428   5.8876   5.5831  -6.8029   1.7317
#>   -3.5755  -1.5488  12.4500 -11.4779   3.9242   0.5745   1.7209 -14.0335
#>    5.4401  -0.6434   2.3415  -1.6133   1.7104   2.6466   2.8619  -5.6875
#>   -4.0917   9.5521   5.5209 -12.8760 -13.2045  26.9454  11.3518  -0.6958
#>   -1.6141   9.6963  -8.7865  -2.3417   6.8534   4.2674   5.2861 -18.4156
#>   -4.4285  -4.4720   2.9013  -2.4991  13.1687 -12.2691  10.4197  -0.5261
#>    1.6306  -6.0883   8.1606  -9.4261  12.9507   9.5293 -10.3306   3.3280
#>   -8.8778  -3.9632  -5.2687   5.1361  -7.5796  19.4838  11.2284  -1.2998
#>    1.2876  -1.1268  -3.5722  -8.0370   8.7132 -11.7825  -9.5280   0.3141
#>   -3.7646  -0.9809  -3.4914  -1.3017  12.1834 -12.2094  -0.6680  -6.3553
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]