Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.6422   2.7924  -5.2445  -1.0746  -6.1598  11.7670  -3.8420  -2.6785
#>   -3.7561  -1.2258  -7.9151  -1.7180   2.1011   2.1352  -8.4484 -15.0835
#>    0.5085   3.9533  -2.4937  -9.9890  -5.6187  18.9932  -4.3046  -7.6362
#>   -1.8687  -5.0245  -1.1614  -7.3313  -2.9345  -3.7011  -1.1842 -12.5681
#>   -0.5714   6.3876   2.1877   2.8623 -14.9784  14.2080  -1.0606  -4.0793
#>   -0.8111  -2.9609  -1.5912  -2.1794   6.0250   0.0042 -24.3139  -2.1512
#>    4.7769  -1.3475   2.3316 -12.5766  -3.3183   8.4968  -8.7246 -11.7900
#>   -4.9863  -9.6916  -0.6596   6.5731   0.1772  -0.4093 -17.8547  -2.5350
#>    5.0886  -4.4308  -0.0121  -0.5116  -1.1531   6.1340  12.3732 -13.3938
#>   -2.5714   6.7988  -0.6296  -2.7975  -2.4728   1.6717   8.9374  -6.0541
#>   -1.5572  -2.7275  -5.7855   2.6663  -5.9592   0.0489   0.2954   2.1139
#>   -0.3855   2.1847  -8.8007   4.2637   0.0113  -0.5055 -13.6081 -15.0998
#>   -5.5655  -6.7756  -7.5635  16.7850  13.5473   9.6970  -9.2147 -17.6331
#>   -4.1929   5.4890   5.7412 -14.3365  -7.9074  -2.7301  18.0999  -2.4298
#>    0.3391  -4.0782   5.5723   4.9290  -2.1955  -3.2209   5.6586   2.7294
#>    0.3253   1.1619  -0.0349 -14.0138   0.7833  -1.7046  -5.6214 -11.5253
#>    2.9190  -7.6368  -3.9634 -14.0984   0.1226  11.0000   2.1030  -8.0923
#>    3.0521  -4.0114   0.2313   9.0539  -4.4189 -26.3621  -7.0620  26.9390
#>    2.2684  -3.0393 -10.5671  -4.4775  10.1830  -3.0298   2.1233  13.6897
#>   -0.0570   5.6555  -1.8907   6.2091  -4.1030   9.1168  -7.6201  -4.1200
#>   -0.4030   4.7369   5.0519  -9.7827 -19.4837   0.3419  10.7008  -2.9806
#>   -8.6631   5.1472   2.5293   2.0277   2.0563  -6.3060   8.2693 -26.5470
#>    2.8047   1.1069   8.3014  -4.7619  -3.5827  -0.4522   6.4595   0.1573
#>    3.0216  -7.6038   0.7406  -2.4424  11.8614  -2.5529   4.0128  -7.7152
#>    3.3145  -1.6718  -5.1789  15.1994  14.7669  -4.1295 -11.7649 -11.9632
#>   -0.0083   5.1825  -1.1882  -4.9004   4.4846  17.5432   6.7651  -9.9113
#>   -2.4879  -2.6639   4.2994 -10.6097   8.8233   0.0206  10.5822  -2.2018
#>   -8.3399   5.3780   4.2747  -2.4799   2.6834  -0.9692  -7.4275  -0.6742
#>    3.9677  -3.6948   5.8200  -7.3888   7.0766   6.6393  -5.1932  -4.2294
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]