Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -8.6867  -0.0503  13.5142  -4.2941 -13.2557   4.7508  -7.2302  -5.8041
#>    3.1262  11.8880 -11.1372 -13.4199 -15.0384  -0.4302  12.5970  -9.2997
#>  -13.8897  10.2481   1.6408   2.2882 -26.6287   6.7141   7.6691   2.8565
#>   -4.3717   1.6723  -3.6682  -4.8041  -4.2435  12.3884   2.9886   3.5974
#>    7.1784  -1.1234   5.4434  -2.0001 -13.0991  -5.0142  -4.3764  -3.9683
#>   -3.5336  -0.0940  -1.3962   1.2221  -5.4040   3.8499  13.0142  -0.2876
#>  -12.0752   0.4050  -0.5337  -9.2639  -8.2126   1.2670  -3.8585   4.2424
#>   -3.6207  -0.9131  -5.9975  -2.4328   9.1095  -5.5047  -9.2628   1.1249
#>   -3.7621   1.9228  -1.3416  -0.5743  -0.3761   0.4834   4.4917   1.7310
#>    5.6021   6.2380  -6.3349  10.3436  -6.8043   7.3048   9.2632   9.9251
#>   -7.2469  16.1261  11.0892 -14.0113   5.7106   3.3881   2.8270  -6.7554
#>  -13.5471  -3.9328  -2.3695  -2.4190   4.3043   5.2635   2.4476 -15.3531
#>    3.4608   3.4454  -9.0681   2.4581 -11.8296   3.1953  -2.8256 -10.8989
#>    5.8406  -2.3147   3.3598   7.4690   4.2544   3.7354  -6.8472   7.3672
#>    5.7313   2.7987   0.1416   6.0178  -8.3849  -8.5804   0.0062  -2.2043
#>   -1.9952  -0.0421   2.7046 -11.7336  -5.9348 -15.7910  -3.8164  -6.2899
#>   -3.2079 -14.2840   1.9151   6.6475   3.1258   0.5761  -7.3537   6.0359
#>   -1.1826  -1.2198   6.3524  14.1255   2.4440  -1.6166  -5.4311  11.7349
#>    1.1976 -10.5607  -7.1565   6.3109  -7.5561 -10.1117   6.7903  19.1275
#>   -3.2064   0.4384   0.9214 -14.1123   8.7626  -3.5258  -7.3903  -9.6812
#>   -6.6400  -3.5876   7.5005   0.5138   0.3058  -4.2568  -2.6867   8.7660
#>   -3.1954  -4.0969   7.1661  -4.1009   5.9616   1.7777   7.7107  -0.4364
#>    2.2177   3.3536   3.8558  -0.1876 -12.5922  -3.3924  -3.5782   4.9435
#>   -7.8703  -0.9285   1.4017  -1.6087 -12.0604  -9.4772  -0.4659  -0.8360
#>    2.6421  -0.2868 -15.4343   8.1299  14.6656  -4.4955   5.0134   6.5886
#>    9.4647  -3.8545  -5.5880   6.5927  -3.3092  -1.8837  -5.1138  -2.7895
#>   -0.6817  -2.3824  -2.7927   1.8499 -16.2706   1.8889   0.4027   1.2175
#>    1.1534   6.9910  -5.2639  -8.3768  -7.7870  19.3899 -10.8954   4.6372
#>   -4.2319  -7.8821  -3.5522   2.7381   6.5012  -1.4675  -9.8912 -25.2016
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]