Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.6547  -5.4597  11.3662   1.4891  -5.6638 -11.3990  10.3785 -12.2572
#>   -5.2737  12.1049 -11.1426 -21.8127  -6.4912  -1.8951  -8.2799   2.2236
#>    2.4319   5.2861  11.3621  -4.5386  16.6069  -6.3963  -9.6982   1.9445
#>    2.4036   3.2287  -5.7853 -11.8675  -2.9683   6.6751  -6.6366 -18.2023
#>   -5.1094  -4.1819  -7.6984  -2.8647   4.8379 -13.5601 -13.6227  12.2559
#>   -3.5288  -0.0662  -4.6908 -12.3634  -7.6508   6.1336  -8.7729  -2.5373
#>    3.1111   1.9531  -9.8054  12.3833 -11.7645   4.5137   2.7507 -10.0450
#>    5.2513   4.0277  -6.3988   9.7328   9.0240  -2.5527   3.0946   5.1641
#>   -6.3348  -4.3065  -1.6504   2.2352 -10.6366   8.2283   3.6260  -3.5249
#>   -3.0750   2.4724   1.1463  -1.1836  -6.4592  10.6023  15.2960  -3.6857
#>    3.6821 -17.6212  13.0013  -6.8021  18.3353 -12.1838  10.5479  -0.0389
#>    0.6373   8.2082  10.1023 -15.5664   2.8004 -17.1291  10.9047 -15.1061
#>    0.1656 -10.8750   0.7038  -6.5855   9.1201   0.1424  16.2241  -2.4898
#>    0.0179  -3.3191   5.6866  19.2046  -9.9267   7.4214  12.2237  -4.7667
#>    4.6162   7.0575   6.8186  13.4308  19.0778   3.6393   1.8333  -5.2680
#>    1.4339   4.1601   2.1909   0.8323  13.6129  14.4614 -15.9289  -5.7412
#>   -5.0725   5.5461   8.9722 -15.3886  11.2898 -15.3429   3.5443  -0.5031
#>    3.0737  -8.1247   0.9163  -3.1141  11.8036  -5.3201  10.2993   2.8328
#>    0.9946  -5.4607  -1.4533 -12.5266  -5.9500 -11.4856  -8.0112   2.1005
#>   -3.0604   9.2960  -1.1421   1.8728  12.0369   1.7284   2.0522  18.0436
#>    0.2940   3.3377  14.9372  -0.3042   1.5348  23.6991  -2.4563   7.8753
#>   -3.7579   4.8194   3.4110  -9.0820  -3.2316 -11.6617 -10.8368  -7.5996
#>   -9.3728   3.9610  -9.9975  -2.0756  -7.1202   0.2236  -6.7141  -3.3324
#>    2.4636  -5.5534   7.0648   6.8724  -1.8453  -3.5832  -1.5622   7.2301
#>    1.0673  14.6355  -9.9130   0.1318  -0.0404   5.2692   1.2678   3.1115
#>   -7.8446   5.5108 -13.0323  16.2700  -7.5881 -11.9766  -6.0894   9.0764
#>   -0.0680   2.8148  18.1885   7.0523   9.0324   1.7933   4.8252  -5.4561
#>   -1.5380  -6.0340   6.7325  -6.9267 -12.2748  -3.1566  -3.1084 -21.9326
#>    0.1443  -8.9633   5.1383  -1.3855  -9.2591 -12.0832   7.9167  -0.7600
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]