Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.9234  -7.2806   9.4841   5.2656  -4.1631  -7.5479  -9.1135   5.8547
#>    6.7251   3.8817  -1.3362   2.5977   0.9214  -6.2308  15.8241  -0.8781
#>   -2.4603   0.4634   4.1179   4.7448   2.7985   1.5243   0.4888  15.5770
#>    0.4267  -0.7224   0.0061  -1.1297  14.5613   4.4218   0.1307  -9.4013
#>    0.1833  -5.7686  11.1619   3.6650   6.2817   3.9791   8.3859  -0.2335
#>    1.3214   6.2052   9.7643 -11.7601  -5.0660  -4.8231   8.8813  11.1443
#>    3.2137   6.0294 -13.3862   3.8929  -0.2893   1.8032  12.9331  15.6461
#>    1.8617  -1.4645  -1.0458   4.4261   7.8358   9.1079   3.1766  -5.8477
#>    1.7002  -4.6037   9.0963  14.0484   0.1291  13.7271 -11.9398 -14.1623
#>   -1.9847  -7.7887   1.7267   5.7430  -1.2683  -5.3557   3.2082   6.4200
#>   -1.3901  -0.7906   7.0917  13.3084   9.2962   0.0522  -1.5405  -0.7853
#>   -0.6564  -9.5817  -3.4979   1.3919 -16.7169   3.8634  -0.9446   7.6568
#>    1.1056  -8.1688  -1.5817 -12.3991   0.9541   1.0334   4.1893   6.6047
#>   -4.4015   1.3324   0.6677 -16.6558   0.4950 -14.8438  -9.8484 -13.3619
#>    0.9216   7.8381  -7.0980   5.9095  11.7501   3.2189 -12.7808  -7.4955
#>   -6.8357  -5.5775  -7.2691 -12.2555 -15.2402   5.6523   5.2473  -7.1518
#>    1.7273   0.5525   3.6988   0.5225   5.1939  -3.4717  -1.4779   6.3422
#>   -3.5475  -4.2411   2.1811  -3.5097   9.5064  -1.2703  -4.7568 -12.8979
#>   -0.6605   0.2464  -3.5989  -1.3860  10.8535   0.3595 -19.3124   2.4627
#>   -3.5668   7.7808 -14.0608  -5.7888 -17.7978  14.9126  10.4147  -3.6979
#>    0.1220  -3.9894   2.1837   1.2860  -4.2341   5.2764  -9.6178   5.1710
#>   12.0515  -1.5343   4.9232  -8.9793  -0.2974  -5.0300   0.3548  -5.5407
#>   -4.5864  -1.0543   2.0440  -4.4689  14.3605  -4.0543   1.1839  -3.7918
#>    4.0475  -2.2675  -4.4756  -7.1519  -5.7331  -6.7579   1.7906   2.2219
#>   -1.6037  -2.2171   5.3104   9.1659   0.7954  18.0858  -0.4819  -0.9980
#>    4.0181   0.2366   3.3160  -1.4543  -2.1106   1.2161 -11.0400  -7.3614
#>   -4.6399  -7.9887  -7.8529  -7.9901  -0.1163   7.1619  14.2008  -4.4655
#>   -0.4701   1.3907  -3.9165  -7.6086  -5.1391  -6.6722  -8.7270   1.5376
#>    0.8263   2.0199  -0.3443  -3.4956  14.8260  -3.3319   0.4458  -4.1993
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]