Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.2571  -2.8375   3.0275 -15.0874   1.3235   0.1913   3.3347 -11.3620
#>    0.2393   9.3840   5.9469 -12.0852   1.7433  -5.7072 -16.5887  22.6444
#>   10.9700   2.6430  -1.9248  11.9617   8.1068 -11.5786  -0.9858  -4.2221
#>   -3.2114   2.2619   1.2575  -0.2386   7.0735 -13.3114  11.4126   8.9763
#>    1.6007  -2.0231   4.4937   6.4341  -9.8340  -3.1746  11.8781 -19.2681
#>    9.5151  -1.7231   2.3247   1.6664  -1.0251   2.8796  -0.3368 -14.4379
#>    2.4506  -8.5462   3.3400  -1.7493   0.4641   2.6318 -10.9493   5.0131
#>   -0.7376   4.6777  -9.6799 -14.6330   5.9224  -3.1364 -11.2936   2.1816
#>    0.3267  -2.3732 -14.9662  -5.1133   4.3651   8.6378  -5.6448 -17.8085
#>    5.8885   0.2830   3.0217  13.2488  -4.1225 -27.9392   2.8967   6.1749
#>    0.8679   4.0503  -0.6321   0.5595  -1.7734   3.5887  -8.0482   1.8105
#>  -11.0753  -0.1174  -1.4324   3.3993 -18.2896   5.4972   5.4781   1.5240
#>   -0.7965  13.0675  -0.3192  -2.9341   4.0633  -2.5099  11.1699  -4.2997
#>    1.0880   0.5017   2.9747   5.5607   4.1540  -3.0392  -1.0147   7.9554
#>    5.7672   4.2134   2.4008   3.4886   1.9007   0.9267   0.8732 -11.9741
#>   -8.5722   6.9330  -4.7572  -0.5619   0.6731  22.2420  12.9583  -5.1037
#>    0.3150 -12.9968  -1.8360  -1.1850   3.3185  -3.0946   7.0031   1.2972
#>   -5.5224 -13.7795  -5.8111  11.7751 -20.4337 -15.4730  -0.2575  11.0466
#>    2.1850   3.9506   2.3771   2.6677   3.2981   8.6332  -6.5184  12.6802
#>    0.7091  -6.2638   0.5250   4.4231   4.2668   2.7757  -5.0757  12.2622
#>    5.1904  -4.4347  16.1715   9.1450   3.8610  -2.0376   7.0037  26.2743
#>   -3.5573   7.0116  -1.8493   2.3225 -30.8692   2.3344   1.7453   1.3345
#>    0.1684  -4.1248   6.4724  -7.1404  -9.1474 -11.0834   7.4233   5.9115
#>    2.6641   5.3568 -10.3765  -3.8865 -10.6351  16.7465   1.0649  -6.1903
#>    5.3541   0.7824  -6.3529   0.5039  -1.1245  -9.3725   4.0015   3.7067
#>    1.4787   7.5620   0.7648  -0.4869   8.3231  -0.8049  -3.2296  -7.3759
#>   -6.6332  -2.1551  -6.3822 -16.7347  18.0051 -22.7084  -1.2105  11.4172
#>    2.2581  -2.5008   2.1784  19.5139 -14.6745  -9.5155   4.0640  13.7698
#>    7.6130   1.2043  18.6044  -2.4436   4.7524   7.3389  16.3337   4.0555
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]