Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -9.3224e-01  1.9959e+00  2.3848e+00  8.6716e-02  1.2238e+01  6.0039e+00
#>  -3.4611e+00 -6.0874e+00  6.0400e+00  4.1266e+00  2.6864e+00  1.2239e+01
#>  -3.9321e+00 -1.4360e+00 -6.3409e+00  5.4203e+00 -3.0380e-01 -1.6147e+00
#>  -5.8953e+00 -5.4831e-01 -1.3274e+00  9.1095e+00 -2.7476e+00 -1.0390e+01
#>   1.0362e+00 -1.0287e+01  1.0963e+00  4.8236e+00  2.0094e+00 -1.2383e+01
#>  -1.5338e+00 -5.5807e+00  3.1850e+00 -1.0794e+01  1.5478e+00 -8.5988e+00
#>  -4.5685e+00  5.6668e+00  3.7085e+00  2.5341e-01 -3.8537e+00 -2.4351e+01
#>  -7.2601e+00 -4.2975e+00 -4.5551e+00  1.7144e+01 -6.2837e+00 -4.2550e+00
#>  -4.7419e+00 -4.0690e+00  1.1835e-01  1.5230e+00  8.3965e+00 -3.1364e+00
#>  -2.0348e+00  3.3886e+00  4.1892e+00 -1.6305e+00 -9.3577e+00  9.1033e-01
#>   5.1930e+00 -2.4307e+00  3.7216e+00 -3.1464e+00 -7.4632e+00 -5.2628e+00
#>   3.9202e+00  3.8260e+00 -5.1402e+00 -1.0402e+01 -1.6305e+00  9.7050e-01
#>  -3.0671e+00  5.7251e+00  6.9542e+00  3.9722e+00 -2.9572e+00 -1.0235e+01
#>   2.9852e+00 -8.9888e-01  2.1673e+00 -5.3547e+00 -2.1282e+00 -1.2688e+00
#>   3.5540e+00 -3.6646e+00 -5.4178e+00 -4.7512e+00  3.3309e+00 -9.3770e-01
#>  -6.0340e+00 -5.3852e+00 -5.1386e+00  8.7557e+00 -9.5567e-01 -8.4686e+00
#>  -2.4003e+00 -1.1303e+00 -1.9837e+00  1.2652e+01 -2.9152e+00 -1.2245e+01
#>   7.1568e+00  1.6707e+00 -9.1672e-01  3.3633e+00  8.1383e+00  6.5145e+00
#>   1.8941e+00  4.5725e+00  5.8898e-01 -1.3634e+01  1.0578e+01  1.6304e+01
#>   1.7525e+00  1.9909e+00 -6.3672e+00  1.7606e+00 -1.0339e+01  4.5358e+00
#>  -5.2920e+00 -7.8174e+00  7.7432e+00  1.2277e+01 -4.2827e+00  1.2211e+01
#>  -3.2385e+00  2.4606e-01 -1.5491e+01  2.3443e+01  8.6799e+00  7.0165e+00
#>  -4.4970e+00  8.3095e+00  5.6369e+00 -1.2152e+01 -1.2498e+01  4.5997e+00
#>  -1.1086e+00 -1.4665e+01 -3.4749e+00 -7.4560e-01  5.2708e+00 -1.0206e+01
#>   8.1130e+00  1.0627e+01 -2.6100e-01  9.8707e+00 -1.7460e+01  8.1356e+00
#>   7.6800e+00 -1.3660e+00  6.1892e+00 -2.8995e+00  1.1813e+01  2.6396e-01
#>   2.0452e+00 -3.3163e+00  1.8463e+01  3.4137e+00  1.1008e+01 -1.6415e+00
#>  -3.2783e-01  1.3313e+00 -2.0481e+00  6.1509e+00  2.4349e+01  2.5756e-01
#>  -1.3184e+00  4.5914e+00  9.2247e+00 -1.6352e+01 -6.6283e+00 -3.5893e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]