Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.7425  -0.1140  -3.9292  -0.1371   0.5673  16.1053   1.4189   3.2460
#>   -5.3275   5.9378   5.8241  -5.1763   7.4594   0.8162   4.0511   4.3798
#>    2.2974  -6.4189   6.8200  -8.8303   1.8136   1.4287 -14.5390  12.4834
#>    0.0047   1.5154   8.7210  15.3092   3.8384   7.7287   3.5345 -12.0092
#>   -4.1305   4.2522 -12.7505   1.6656   9.0682  -5.9529  -4.3489  -4.9378
#>    6.2679   5.3866   0.3161  15.4660   6.7557   1.0186  -6.8452  13.8601
#>    0.7260   4.5586  -0.6076  -7.1991   0.6636  10.5773   3.5951   4.1094
#>    2.8855  18.8059   7.4180   5.6695   5.3975   5.8358  -0.2465   6.1338
#>    1.3751  -3.7199  10.6373  13.2313  22.5241   0.7659  -6.3265  -2.4687
#>   -2.0034  -5.0810 -15.5668  -4.8708 -15.0944   7.4064  -0.5007  15.4169
#>    1.4397  -7.3979   5.3638  -6.1193  11.7855 -16.2841  13.7513  -5.1697
#>   -0.4689  -5.6136  -6.1342 -10.5604  -7.4025   8.9524  -0.7059  20.1923
#>   -0.1777   0.0821   4.9260  -2.9225   2.6632  10.6563   5.0424  -1.7273
#>   -1.9205  -6.7166   5.5766  -8.9845   0.2756  -2.3865  -1.8392   3.5197
#>   -1.0297   0.3282   4.7804   4.7785  12.3605  20.6644  -5.1044   7.1147
#>    7.0565  -2.7273  -4.0247   8.9343  -3.8258 -14.9105 -14.7229   4.6567
#>    4.5229   0.5925   1.9932  -4.1067   2.2677   1.4839  -8.0333   5.8453
#>    3.2185 -14.1720   7.2177 -15.5749  -7.3815 -16.8526  10.5367  11.0690
#>    5.1170  -0.3350  -8.3603  14.2062  -0.0727  -2.9667  -0.9302  -2.1008
#>    0.3725  -5.7678   0.9078  -0.7164   3.0376   8.5407  12.5427   6.3928
#>   -7.8154   7.9762 -10.1483  -2.8933 -14.1705   2.3893 -22.3592  10.7860
#>    7.1047   4.0130   4.2930  -7.5901  -5.9511  11.8991 -22.1199  17.7319
#>   -4.5709  -2.8618  -8.0598  -1.5742   1.5381  -0.5736  21.8864   4.9246
#>   -1.9708 -11.2157   2.5225  -4.6493   3.4711   2.6460  10.0497   6.4213
#>   -3.8513  -3.6146  -8.6462  -8.5316  14.2756  -1.1116   5.8362  -4.0001
#>   -4.8301   5.5279  11.0843   1.1535  16.0331   0.6202   1.8037  -3.0764
#>    1.6170  -2.3651  -4.7330 -10.6063   3.0534 -16.3598  -0.2254  -2.7894
#>  -11.4576   1.6193  -4.5685   5.2393  -2.4445  14.7896  -2.3955  -2.6836
#>    6.9892  -4.0574  13.9968   7.1619  -8.6467   3.8835   4.3811   6.9211
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]