Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.6012  -3.2203   8.2308  13.7167  -7.1236   6.6328  10.8817  -3.1470
#>    2.2065   4.1915  -0.4441  19.3433  10.4358   1.8002  13.3724 -12.5769
#>    6.5026   0.0677  17.3140   4.2231  -2.2690  10.2874  23.1514   7.5521
#>   -3.6353  -0.7161  -8.4999 -16.2789  -5.8376  -6.7906 -16.0248  16.1606
#>    0.1535   2.6930  14.8432  -1.1421  10.3787   8.2214  -8.7325   9.1599
#>   -1.2362   2.3819  -5.7836 -11.4007   2.4461   6.5865   3.1976  -5.9758
#>   -3.4332  -1.0416  12.9814  13.5432  -1.6029  -2.5528  -3.9980   1.4831
#>   -0.9593  -2.0680  16.0455   4.9989 -17.5192  15.9101  -1.6578   7.3829
#>   -2.0865  15.1047  -2.5847   1.5254  16.3457  -5.6239  -3.7079  -4.4817
#>    2.7314  -4.1967  12.2699   8.1948 -11.7533   3.1718   5.3235   3.9010
#>   -2.7443   0.1883 -12.3145  11.5945  -6.3745   2.1454  -0.2788   8.7322
#>    4.5710   2.8295   3.0259  10.5340   2.6589  -0.8868   9.6249 -17.1640
#>    1.9136   0.8078   5.7384  22.2803  21.5675  23.1109   1.2104  10.0757
#>    0.1284  -2.1085 -13.8320  -0.0379  -4.1168  -2.3630  -5.3981 -15.0278
#>  -10.0312   4.9617   7.9166  -0.8348   4.3428 -16.1889  -9.7605  -0.1004
#>  -11.3965  12.9161  -7.4938  -2.7577  13.4203  -4.7948  17.0263  -8.2036
#>    0.1695  -0.4118  -7.7046   6.1718   8.5808  -0.7251   0.8679   0.3007
#>   -5.2856  -1.7000  14.0001   2.6080   0.4657  18.5183   5.9300   3.1704
#>   -1.8787  -1.0729  -0.3855  -9.3791  -5.4064   7.1918 -11.2657  -3.2791
#>   -2.4123  12.0163  17.8684 -10.5981   6.4290  -0.5395  -8.7724  -2.2357
#>    1.7126  -6.4478  -1.0476  11.2582  -4.2978  -1.9348  10.6632  -2.8129
#>   -0.3325   3.6258  -0.4160  16.8279   6.7264  12.3276  17.5635 -14.7457
#>    4.0322  -5.6933  15.4201  12.4205 -11.8602  -3.9549 -10.2792  -1.3215
#>    2.2177  -3.4420  -3.0154  10.6720  -5.5626   0.6513  -7.9614   6.8003
#>    1.1791   0.8012 -18.9799   5.9601  10.4750  -7.1448  15.0576  -1.5701
#>    2.3859   8.1546  -7.8806   9.9083  -0.6572   4.6550  -0.3882 -10.3260
#>    7.0098  -3.7357  -2.7588  10.7364   0.6185   1.9782   0.1851  -6.3515
#>   -7.6676   1.5563   4.8748   9.7903   7.1687 -10.3581  -7.6672  -2.6947
#>    1.4229  -1.5663  -6.8532  -4.7778   4.9298  -1.4206  17.1215  -2.5845
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]