Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.6851  -7.4879   7.8627   2.1011   6.1813   6.1876  -6.7604   5.8731
#>   -6.9953   7.2242   4.9340   3.3770  -3.7243 -10.3868 -12.4656  -8.2818
#>    6.7513  -2.6709   5.0128  16.8339 -10.2127  -8.6403  -6.8349 -10.0156
#>    2.7125  -4.9483   4.5488  -1.8358   8.8843  -9.6633  17.3118   6.7950
#>    2.8805  -2.4819  -9.1845   1.9963 -10.1163  10.1048   6.6256  -2.7318
#>   -6.7226   4.0831  -3.0329  -7.8906   8.7028 -14.5036  -6.5115  -8.6078
#>   -4.4438   8.6076  -1.8594   4.1283  12.0067  -0.4217  -2.6040   8.7201
#>   -4.5501   2.0104   2.4446   0.3526  -3.9713  -4.5672   6.9476  -3.7014
#>    0.1162   4.2999  -4.8790  -0.0527 -11.0151  -6.3885  -5.2040   6.8037
#>   -2.9453   6.0640   1.2021  -6.7875  -5.2460  -5.4450  -5.0200   4.2473
#>    2.0360  -5.3232  -6.9087  -0.8072   6.2786   2.6644   7.2203   6.2317
#>   -6.3820   0.8968  -8.8117   0.4407   1.7556  -0.3249  -0.0429  -4.9799
#>    2.5674   3.6517   0.8179   3.0492   6.3819  11.2050   4.0867   6.2484
#>    1.8109   5.8450   6.9508   3.6347   2.0986   2.3536  -4.5000   6.5310
#>    1.7800  -4.1873   7.4447   5.7679  12.3441   8.0623   3.2410   3.7636
#>    2.0306   0.5260   5.6710  -7.2246  -2.3128  -8.7923   2.0624   3.5153
#>    1.0936   1.0248  -6.5713   4.7590  -5.1891  10.5827  -2.0666   4.1467
#>   -3.3883   0.7396  -0.1437  -1.6672   4.6814   9.3406  11.3223   1.0397
#>    0.1236   1.5507  -7.8929   7.9607   1.7498   0.6727   2.6967  -2.4959
#>   -6.7613   3.7849   1.8982  -1.3070  -2.1107  -1.8209  -1.2315  -1.3782
#>    4.7972  -0.6533 -12.3384   1.8838  -3.1318  -5.3857   5.0314  -4.4800
#>   -1.0249   2.8525   3.4276   4.4104  -1.0310  -0.7328  -3.3431  -7.2878
#>   -0.0380   1.8662  12.8814   1.4060  14.6396   0.2668   3.1171   0.5370
#>    2.9337  -2.9946   4.8681  -7.1598  -0.3958  -7.4334   3.7319  -0.5605
#>    1.3820   4.3551   4.9861  -8.3421  -3.0264   0.2906   0.7973   9.2900
#>   -1.9313  -0.2406  14.1263   1.5275   4.3621  -9.1830  -1.2445   0.5265
#>   -2.2972   0.7224  -5.0229 -10.2066  18.6383  12.4550   0.5041  -3.6972
#>    0.6422  -1.4065  -4.3360  11.4841   1.6873  -3.0108  -7.7710  -2.0095
#>    1.2264   9.5499  -0.3652   7.9479  -3.1966   0.3778   9.8269   2.2679
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]