Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.6830   2.9650  -5.5382   8.9186  11.2195   0.8629   2.6964   9.4712
#>   -1.7070   4.3492   8.0043  10.6571   0.3601   2.4528 -12.0436  -7.6047
#>   -0.9914  -2.8566   2.1509  -1.8343  -7.4371  -1.7817  -7.7771  -1.9744
#>   -3.9998   0.7933   0.1155   0.1321   0.5810   1.4915  10.0938  -4.6731
#>   -1.7308  -4.7594  -6.7404   2.3765  -3.8116   8.0808   5.8106  -7.0818
#>   -2.2626  10.9743   6.9127  10.6143 -15.0723  -7.1755   1.2079  -6.2266
#>   -1.0185  -3.3680  -0.5049  -3.6329   7.4979   5.6949   0.8594  20.5962
#>    3.9791   4.8527   5.0225  -6.0007   3.2177   4.1591  -5.4839   0.1296
#>    1.5124   1.4859   1.5521   2.4251  -3.0096   9.3336 -17.7847   4.7096
#>    1.6333  -7.3563   4.0900   5.8158  -7.1717   9.2768   1.1961 -12.5629
#>   -1.6897   4.1147  -1.9799  -2.5688  -1.2641   3.5820  -7.5374  -3.4959
#>   -3.2444  -5.5845  -4.9612  -1.2565  -8.0886  -3.8759   4.4811 -11.4880
#>   -2.3904   7.2898  -4.3013  -0.5552   0.2981  -3.7669  -6.3700   7.2224
#>    2.0232   0.6971  -6.0640  -0.9484  14.5916  -3.9507  17.9024  20.4342
#>   -1.7923   2.1130  -9.6393  -9.6266   5.8873  -1.5971   8.6634   4.0906
#>   -2.2084   3.9594   3.8908   7.1670  -9.4340  -3.7885 -12.2797  -0.8751
#>   -6.0269  -2.7471  -6.3257   0.5263  -5.6403   2.9769   2.4260  -7.8800
#>    0.0325   2.0510   3.5840   5.5386   2.7608  -2.4625 -10.9930  14.0415
#>   -0.0833  -1.3747  -1.1258 -13.0676   0.4440 -18.8078   7.5367 -11.6033
#>   -3.8646  -8.8496   2.8317  -6.8369  -9.3832 -17.3035  -0.4740  10.2151
#>   -0.6951   8.3158   2.0426  -1.0327   4.8292  -2.0160   3.9468  -8.1769
#>   -4.3425   1.5466   2.3380  -1.0927  -4.0691  -1.2096  -4.2089  -9.8286
#>   -0.9145   2.1939   9.0824   6.2803  -2.4124  -2.0090   9.2380 -12.6534
#>    1.6476   5.2511   9.3907   8.4700   7.2479   2.3701  -6.4555  -7.1649
#>   -0.2765   6.1312  -0.7332  -5.1036   8.1794   8.2149   9.3591  -9.0113
#>   -1.5245   6.8402   3.3052  -1.0557  -5.6113  -7.0661   8.5449  -2.7826
#>   -2.5535  -2.7037  -5.9317  -3.4869   3.5218  -0.5531   2.4454   3.4778
#>    3.4907   0.2140   3.4302  -8.4799   1.5092   3.4763   4.6976  14.2595
#>    1.4584   7.8266  -2.9125   1.2510 -12.2742  -2.3780   4.8999  -8.8204
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]