Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.9211   1.3511   2.0683 -13.0644  -6.9527 -10.7137   3.7730  -8.5478
#>   -6.5693  -5.4825   4.0959   9.3739  -1.1509 -11.1555  -6.0745  -5.1503
#>   -6.1888  -8.4051   5.3868  10.8176  13.8304  -4.5643 -11.9518  18.4834
#>    1.0036   0.1932   0.8812   5.2339  14.6172  10.5493   8.0721  -3.0311
#>    1.2876  -5.1590   6.1053   4.6886  11.1610   0.1887  -4.1540  -2.1612
#>   -0.9968   0.6978  -1.8459  -9.6686  -7.2875   5.2570  -5.5048  14.9170
#>    1.1083  -6.4162  -8.6855   5.6385   8.2375   3.2163   7.0482  -6.1981
#>    5.3094   0.5321   5.9592   4.9252   1.2951   2.2461   9.4089  -5.4960
#>    2.8580  -2.5644  -8.3588   0.5246   0.5925   2.8952  -9.8749  -9.8835
#>    2.5943   5.6865  10.2064   7.5028  11.6174  19.0694   7.5511  10.2127
#>   -1.5005  -1.0004  11.7291   1.1653 -10.5907  -7.1627  -3.5667  -6.2047
#>   -7.7215  -3.6644   1.0110   9.3568   2.3130 -12.8325  -1.4332 -14.3041
#>    5.0097   8.4301   2.0691   1.1094  -2.8391   0.7234  14.6303  13.2351
#>    3.5773   6.3869  -8.3432  -9.5571   2.6397  17.0927   5.3649   1.6376
#>    2.1051  -5.0815   1.7187   2.5503  -1.2776  -6.5920   0.6143   6.0772
#>    6.6344   4.6604   0.6654   1.6637  -4.2368   3.4789  16.9188   1.9887
#>   -0.7102  -0.3761  -9.6750  -2.1587   0.5489  12.9033 -14.3575 -10.7983
#>   -6.3651  -0.2174  -4.2774   7.7909 -11.1447 -11.2112   7.0801   2.5944
#>    3.9831  -1.3060  -0.8516  -7.4025 -10.7924   0.9817 -15.5256 -14.5591
#>    1.3255  -4.3780  -4.9212   6.2227   0.5282  -0.2884  -4.8832  -6.6226
#>    4.5917  -0.3275  -1.0628   1.7060   0.3860   1.8325   5.8329  -4.3678
#>   -1.6940  -1.4831   0.8515  -1.0918  -1.2381  -4.2996  -5.4404   4.3136
#>    6.2781   3.0221   5.0934   6.0500  -5.7426  -5.2814  10.4106  -1.4932
#>    4.9426  -1.5315  -5.5811  -2.2502  -6.8306  -9.6872   8.9615  11.1151
#>    4.1493   6.1935  -0.6068  -0.7367  -6.4812  11.9083   1.3716  14.0667
#>    2.9611  -5.8348  -0.2550   0.8047 -16.5453   1.8317  -6.9644 -12.1558
#>   -3.1731  -4.1306  -2.0690  -2.0394  -0.2801   1.0238 -13.9057   0.0265
#>   -1.3434   2.2974  -2.4196  -5.2392  -1.1760  -2.3027  10.7444 -14.5517
#>    1.4976   6.5465   1.7604   2.4022  -8.5534   2.3501   5.6750  -0.0234
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]