Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   8.0520   4.7158   8.9140  -0.4262  -5.7749   2.2878 -10.4537  10.6281
#>    4.7023   0.1372   1.3290   0.2860   8.5461  -2.9366  -1.5405   7.3100
#>    4.9153  -2.3400   3.5097   3.2622   5.9279   1.9859  -9.9406  -3.4375
#>    3.1223  -9.2998   0.1821   3.4526  -5.4449   5.5557  -1.6086  -0.0735
#>    1.9220   6.5467   6.2649 -11.4512  -2.7813  -1.3988   9.5924  -7.6516
#>    7.7494  -0.5448   0.6288  -8.3511   6.9204  -2.8406 -16.5662  -2.5006
#>    5.0004   0.7543  10.6646  -6.8743   6.7831  -6.5080   9.5457  11.2651
#>   -5.5669  -9.6487  -2.0705  13.6616  -3.2039  19.2523 -13.9139   3.9560
#>   -8.3855  -1.1692   6.7825  -8.4058   2.9204  -1.9072 -11.8497  13.2875
#>    1.5538   3.9424   2.0482  -4.7537   3.8269  -0.9719  -6.1889  -4.3419
#>   -2.6438  -7.9851  -0.0598  -8.0771   5.3917  -9.1941   3.2422   1.9509
#>    7.9564  10.5172   4.0967   7.1987  -9.9427  -2.3450   0.4401   1.7693
#>   -0.0457   0.4684   0.3743  -4.0607  -1.1444  -0.6814  -1.1927 -13.8526
#>    0.8792  -6.0093  -2.1718  -1.7350  10.0110   4.9167   8.8506  -6.6238
#>    7.7427  -6.4933  -2.8245  -9.8389   0.1308 -11.3544   4.6251  -9.3225
#>   -5.4021  10.3074  -2.1391  14.4937  10.8596   8.8541 -12.7576  11.9023
#>   -4.9496   3.6615 -13.8201  -7.4383  -6.8935  -3.5014   6.1017  -2.7777
#>    1.7789  -5.2157  -9.8713  12.3214   3.8627  -1.6757  -6.4748   0.7927
#>   -0.9473  -4.7522  20.4131   0.8770  -0.9726  -0.4498 -10.0305  10.5526
#>    1.6510   3.0232   1.0542   1.3903  -6.5638 -13.9531  10.2093  -4.0633
#>    3.3495  -4.7695  16.0527  -5.5948   6.2697  -6.4385   6.8351  11.6516
#>    9.4316  -4.9704  -8.4493  10.0077   1.9695   5.6561   0.1191   2.7991
#>   -0.1595   0.5238  18.3986 -10.8093  22.8761  -3.9393   1.5666   1.8619
#>   -0.3679   3.5998  -6.7402  -0.6707   0.6304  -2.6442  14.3967 -12.7328
#>   -0.7568   0.5892 -10.0568   8.9245  -1.6124  12.7597  -0.7930 -15.9363
#>    1.8244 -11.2025   0.3427  -7.6638   1.9190   0.7749  10.2306  -0.5296
#>    6.4671  -1.6782   0.4100   4.3791  12.3438   1.7855   2.8264  -6.0921
#>    3.0153   4.0947  -5.3500 -15.5991  -4.3496  -4.4980   8.6631 -11.2873
#>   -0.0773   2.3356  12.5095  16.7552   5.2173   0.5886  -6.6109  -3.1752
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]