Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.9806  -5.2695 -10.2718   4.1637  -1.6734   1.7903  -3.0032   1.0966
#>   -1.1051  -7.6132  19.6576   0.2186 -11.2793  -3.1267  -6.1077  11.3742
#>   -2.4800   8.4714  -1.6839   5.4293  -4.7265  -1.6143   7.2012  -4.0170
#>   -3.3689   4.1127  -4.1141  -5.2946   0.8161 -12.8534  -7.7479  -6.6610
#>    4.0037   2.6298 -18.9733  -1.4354  12.5386   7.7923 -15.4220  -3.9129
#>    1.5010  -0.6600 -12.1592   2.4447  11.6108 -20.2343  -9.1467  11.6664
#>    1.5914   1.5170  -4.7928  12.2544  12.2894  -4.0701  -4.9699  -3.9814
#>    2.9846  -8.2546  -9.2085   2.9383  17.4497 -13.5101 -38.8353   9.2243
#>   -2.7863   3.9190   5.2582  -1.9683 -12.3544   3.3401  16.7882 -10.0712
#>    1.6817   0.7294  -4.9636  17.5610  -3.9581  -0.4045   7.1166 -20.3123
#>   -4.3435   0.3748  12.1423   6.4000   0.8197  12.7808   5.6979 -20.2240
#>   -1.3106   1.8286  -4.9973  -2.6947   1.8410   0.5495   1.0101  -0.5120
#>    0.3502  -2.1965  -7.0980   0.3237   2.0190   0.4276   4.7521   7.2375
#>   -2.8233  -1.2819  13.7142   5.1656  -0.6584 -12.5837 -12.9540  -4.1519
#>   -1.9986   1.4978   5.6881   0.9299  -6.5483  -6.5783   0.2173   0.3474
#>   -0.6616   0.2729   1.5751  -0.6422  -4.3630   5.1691  10.6994  -7.4736
#>    0.2373  10.7481 -13.6467   2.4346   4.3071 -10.2650  13.8771  11.1018
#>    3.7981  -5.0476  -0.6253   0.2173  -2.4355  12.8807  -2.9376  21.1926
#>   -1.3038   0.9859  -7.6001   6.7858   3.7978  -8.8111  16.1628   2.7221
#>    4.0013  -4.6338 -12.3224  10.7539  -3.0098 -17.8045   2.3215  12.7284
#>    5.3266  -3.0060  -5.3042  -6.3672  16.7400 -12.3366   4.6881   4.7209
#>   -4.4239   3.7414   9.2690  -6.6596 -11.0613  10.4287  -8.3800   2.8344
#>    2.5749   0.7887  -7.4790   0.2708  11.7160  -0.1781   2.0914  14.6367
#>    1.5968  -2.2695   1.5136   1.8794  -1.6656   0.6957 -14.3309  -6.6874
#>   -0.6627  -5.4695   5.4841  -5.6227  -3.0065   3.1511 -10.0430  -7.5497
#>   -4.1172   5.7796 -13.3937  14.2038  -2.3657   0.1201 -15.0464  -4.4451
#>    2.5669   4.9018 -11.6065  -3.3351  -5.9236   8.7813  -5.9928  -6.1358
#>   -0.3014   0.6991   1.7414  -8.7715  -6.3488  12.3842  -3.8101 -12.4477
#>   -0.0469   2.7470   0.9554  -8.4080  -4.6499  -9.2610  -0.2730  -3.7624
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]