Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -9.5317   6.0858   2.4977   2.1024   4.7119  -6.0658  13.5353  -3.4347
#>   -1.2730   3.4932   6.8790   5.2023  19.1510   3.4952  -6.7958   8.9563
#>   -3.2638   6.8739  -6.9041  -0.3023   1.0388 -11.1189   6.6835  -1.9491
#>    3.9449  -7.1013   6.6537   1.4952   9.1115  14.4857 -11.7158   2.4760
#>    0.3311  -5.8301   0.6301  -2.2589  -4.4077  -2.5847   4.9672   9.6548
#>   -0.7473  -3.5742   5.8439  -7.7619   7.9972   2.8074   0.5779  -1.7672
#>    4.7769  -0.6201   2.9926  -3.5582  -6.1822   3.2284  14.0879  10.4049
#>    2.8279  -0.3586   0.4584 -12.5984 -22.0795   5.6578  -5.4372  21.8473
#>   -2.3645  -1.9843  -0.5584  -6.7227   8.9121  -3.9839 -14.2687  -5.0207
#>   -0.2815   2.3595 -10.4818  -1.1607  -5.1127 -12.9233   1.4015  -1.0878
#>   -4.9664   1.9683   2.2568  -1.4202 -17.4839   7.7744 -14.8552  -3.1574
#>   -1.8947  10.2228  -5.9396  -0.5092  18.3780 -14.7488  -7.0767   2.8174
#>   -0.1426   1.9490  -4.1557   6.4773   1.0485   4.6563   9.4404 -11.7446
#>   -1.0952  -3.7739  -0.4542  -5.7690 -10.6438 -12.1030   3.3822   3.6616
#>   -6.0346   3.7387  -2.6655  11.8425  -0.6803  -0.3909  19.5359   5.9046
#>   -2.0984   1.2234   2.1985 -10.9551 -13.1473  -4.6879   7.5939   5.8184
#>   -1.7001  -7.7013  -4.0305  -0.0247  11.4818   2.4201  -1.9728 -18.1240
#>   -2.3028   6.1543   0.2391   4.6238  -3.1921  -9.8528  -1.6005  -2.1628
#>   -3.2649  -4.7227  -3.5114  -6.9365  -3.6654  -0.0816   1.6070  -0.2986
#>   -1.7305  -1.2148   2.2312 -12.5283   0.8051  -0.3097  -1.8354  10.3808
#>    3.6496  -8.9296   6.4332 -12.7120  -5.5299   3.3201  -4.2175   0.1009
#>   -5.2649   2.6031  -4.8205  -1.6150   4.6447  -7.7336   2.4089   2.4539
#>    3.0164  -6.6747   1.4780  -1.4447 -10.4356   8.0327   3.8766  -3.0747
#>   -5.9746  14.8020   2.4895 -10.7515   9.8194  -9.0172  -3.4040  13.0212
#>   -1.3820   3.3041   0.0682   6.3324 -18.1511   6.3413   6.5460  -1.2025
#>    5.0010   5.7291  11.5521  -0.3299  11.8799  -9.2577 -13.2032   6.9392
#>    4.8948   8.7952  -0.6262   8.2713  -7.2831   4.2532  -0.6774  18.6441
#>    0.5990  -7.0833   2.1201  14.9796   8.0311  16.5751  -0.1868  -4.8497
#>   -3.6714  -1.3036  -0.4202   2.5708   6.7304  -0.0540  -2.5649  -6.0644
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]