Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.1671   7.1783  -9.2886   5.4169  -5.9422   1.7889  -4.0599  -0.6374
#>   -1.1294   0.2957   3.0597 -15.8299  -9.0451   0.6617  -6.7137   3.7968
#>    3.1256  10.7184   9.7659   6.5525  -1.6311   4.9970  -1.1940  -1.3421
#>   -0.0272   1.8206  -7.8545   0.2107  -0.9940   1.0770 -12.5720  -1.6372
#>    0.3761   0.0786  -1.1497   2.5701  -7.4145   9.3564   8.5125   6.9005
#>   -6.2990 -16.5550   6.6367   8.1892  -4.4349  -3.1654 -17.4273  -3.6168
#>    5.2430  -3.2422  -9.0294 -18.7197  13.9652   1.1003  -6.4473  -7.6854
#>   -4.9220  -8.6618  12.6486   2.1842  -0.5857   7.1824   8.9542  22.1628
#>    5.8147   6.5597  -1.7439  -1.7525 -10.4275   8.8624   4.2460   7.6029
#>    4.1444  -1.4289   5.7585  -4.2125  -4.4048   7.3266   3.5108  -3.1331
#>   -4.9483   7.9947   5.3771   8.8952  11.3140  -3.3268   8.4391  -6.7935
#>   -0.5203  -0.6127  -4.7459  -6.5609  -9.8680  15.0034   6.7732   0.9285
#>   -5.0590  -0.1785  -9.0807  -5.3041  -2.0781  -0.8794  13.5924  -1.8926
#>   -7.6822   2.1451   5.2635  10.8806  -5.1891  -2.7949  12.1192   4.2549
#>    8.5217  11.9724 -17.1747  -5.5842   7.5260 -14.9532  -3.5313 -10.0087
#>  -10.2189  -9.4487   2.2789   3.4801  -6.6473  -9.6191   2.0010  12.2604
#>   -1.2974  -2.1635  -7.5180  -2.5186  17.1595   6.5483   8.8341   8.9589
#>   -3.0501  -5.5498   4.3389  -6.4272  -7.8915  -3.4498   7.3404  -0.3165
#>   -0.3393   3.3226  -3.9072   5.4852  -1.2417  -1.0145   8.7341   2.0199
#>    0.2276  -2.4900 -10.7515  -9.5008   7.2306   2.8438  -6.5744   1.8148
#>    6.0097   0.0533   9.7156   9.6563   9.9500  -3.4949 -11.2516  -3.0793
#>   -6.3843  -5.9056   2.6191   2.1581  -1.5270   0.3657  20.2710  16.9459
#>   -3.1244 -11.2591  -7.9871  -7.2412   5.6728  -8.9034   4.0325  -9.0023
#>    5.1941   0.3815   3.2231  19.7607  33.0583 -10.8697  -7.7223  -3.7240
#>    2.2460  16.9138   0.1452 -15.7139   1.3742   0.8085   3.3863 -11.1432
#>   -2.5324   2.1554  17.3230  -4.3853  -2.6764  20.2694   9.2610  12.0836
#>   -1.0378   3.8796  10.5330   7.9518  13.5178   3.9517  -0.3999   6.0871
#>   -4.4331  -2.6390  -4.7733  15.1895  -0.7641  -3.5937  -6.8475  -4.9697
#>    4.6615   5.5079 -13.6573  10.4184  -0.8195   1.6170  -5.9361  16.7004
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]