Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.1366   3.9875   2.4968   2.8538   0.0094  -3.7842   4.9170  -8.9624
#>   -0.3823   0.6851   4.9810   1.8474   0.8745  -2.3809   0.5494   4.8103
#>    5.1476  10.3808 -12.0336  -7.0902  13.9435   5.4523   1.0459   6.0676
#>   -5.2466 -10.1511  -3.0988  -0.9101   2.3438   5.4581  -4.9184   2.9138
#>   -3.8040   5.6768  12.4575 -11.4876   6.4924  -7.4172  -4.2757   3.7413
#>    4.7127   5.4282   3.2867   4.7253  -7.0874   6.4290  -7.8179   8.7026
#>   -2.5910   5.1621  -0.9587  -1.6682   4.4982  -9.8023   8.2377   9.2338
#>    0.4782  -8.9193  10.7938  -8.2887 -13.3779  -7.1019  -1.6222  -7.6646
#>   -1.0660   4.2281   1.7523 -15.9563  10.6577  -8.5994  14.2270   8.4738
#>   -1.1537   2.1560   3.4602  -3.0727   3.8624   4.4272  -6.0987 -10.6693
#>   -4.9063  -0.1034   8.4965 -10.5345  -3.8857  11.0228  -3.5469   0.2733
#>    2.6441  -4.4088  -0.9838  -3.6544  -6.8433  -3.1652   2.2491   7.4611
#>    1.4074  -2.6279  -1.3804  -4.8397   9.1622  -0.8016 -11.3756  13.4609
#>   -0.3653  -9.5083   4.4878  12.6131  -7.9264  -5.9694  12.0710  17.0901
#>    1.2854  -1.0798  -1.1858  -0.5647  15.5319   3.9501  10.5481  -3.7932
#>    9.2443   6.0029  -2.9194  -4.4586   2.2951  -1.1318   8.4747   4.5542
#>   -4.0673   2.2072  -7.6414  13.4231  -3.7177 -10.6171   6.8721  -7.0747
#>   -2.8336   3.6525   9.1218  -6.7268 -10.3230  -6.4753  -1.1212  -6.9193
#>   -0.2861  -2.9760  11.0878   9.5421  -0.4973  -0.4647  -7.4589   0.8508
#>   -3.6835  -3.9664  -4.2988   3.0831   6.2622  -6.0228  -0.2186   1.1776
#>   -6.3303  -1.5710  -2.5018  -2.8768  -3.8353  -1.8763  -1.6023   5.0652
#>   -5.2048   4.9755   1.1503  11.3289 -13.7027  -8.7385  -5.9131  -4.7231
#>   13.2867  -7.2212   2.5480   1.6154  -7.6214   7.3005  -4.9875 -10.5827
#>   -4.7821   2.8892  -5.8312   1.7638  11.2977   2.8087  -0.7314   0.9318
#>    1.9093 -11.5290  10.7354  -7.8191   0.8473   5.3843  11.3820  -3.2594
#>  -10.5864 -10.4527   9.5450   9.4524 -15.0834   5.7240   7.1540  -3.4333
#>   -4.9595  -5.0304  -0.2835   1.0785   6.8152   3.2915   1.9289  -2.0166
#>   -1.6831  -4.1636  -3.8987  -7.2591  16.1817  -6.1284   0.3124  -5.3968
#>    6.8116  -3.8727  -9.6699  10.1883  -2.2544  -4.4434  -8.5082  -3.2169
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]