Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -2.3471e+00 -5.5214e+00 -7.1883e+00 -3.3391e+00  1.1676e+01  5.6366e+00
#>   7.3674e+00 -6.4228e+00  1.7511e+00  4.6599e+00  1.1116e+01 -1.9410e+01
#>   3.5104e+00  8.5132e+00 -5.7019e+00  7.0699e+00 -3.5591e+00  1.8423e+00
#>   1.7652e+00 -1.4667e+00 -4.3814e+00  3.1213e+00 -1.5195e+01  2.2573e+00
#>  -2.7572e+00  2.4448e+00  8.4316e+00  5.9434e+00 -1.0595e-01 -1.6845e+01
#>   1.7516e+00  1.8575e+00  8.3255e+00 -2.0144e+00  1.0139e+01  1.2775e+01
#>   5.5142e+00 -6.3848e+00 -6.5034e+00  1.1401e+01 -3.7556e+00  4.4350e-01
#>  -1.3314e+00  6.1014e+00  8.6935e+00  1.9661e+00  1.0139e+00  1.1139e+01
#>   2.9675e-01 -2.8930e+00 -4.8591e-01  4.8330e+00 -1.6654e+00  1.4573e+01
#>  -2.0264e+00 -9.1697e+00  5.1173e+00  1.1395e+01 -1.3013e+01 -7.2432e+00
#>  -7.0097e-01  5.1242e+00 -2.5765e+00  3.0450e+00 -4.2029e+00 -3.7335e+00
#>  -1.1243e-02  2.1744e-01  3.3394e+00  9.1446e+00 -1.4259e+01 -9.2852e+00
#>  -2.5335e+00 -5.1227e+00  3.6579e+00  1.9029e+01  1.0125e-02  9.0338e-01
#>   3.8419e+00 -3.2488e+00  2.1928e+00 -9.8731e-01 -7.9473e+00  4.5066e-01
#>  -5.9219e-01  2.4652e+00  1.3516e+00  1.5391e+00 -4.2912e+00 -2.4629e+00
#>  -2.2368e+00 -1.8955e+00  7.6902e+00 -9.2335e-01 -1.1435e+01 -1.1756e+01
#>  -4.5276e+00 -9.0118e+00  6.7973e+00  2.3331e+00  4.8413e-01  1.3767e+01
#>   3.1612e+00  7.0814e+00  6.3858e+00  9.2199e+00  2.9335e+00  3.7323e+00
#>   3.0356e-01  2.9994e+00  8.0049e+00 -2.7098e+00  7.3966e+00 -1.1528e+01
#>  -3.7839e+00  1.7456e+00  1.5813e+01  1.6760e+00  1.3070e+01 -1.7835e+01
#>  -1.4838e+00  4.8363e+00 -5.1260e+00  7.6283e+00 -5.9534e+00 -3.8509e+00
#>   3.6361e+00  6.3061e+00 -1.0842e+01 -6.5910e+00 -3.3282e+00 -6.8478e-01
#>   1.1014e-01 -2.9598e+00 -1.3491e+00 -3.2757e+00 -1.8681e+00 -1.1597e+01
#>  -1.1385e+00 -1.0596e+01 -6.6116e+00 -1.6001e+00  1.2335e+01 -1.0625e+00
#>  -3.2653e+00 -8.7581e+00  3.1779e+00 -7.9638e+00 -1.0518e+01  1.4649e+01
#>  -1.8853e+00  4.6159e+00  1.0118e+01 -3.6869e+00  6.2981e+00 -2.8592e+01
#>  -9.8523e-01 -2.0836e+00  5.5839e+00  1.1389e+01 -7.5982e+00  1.4064e+01
#>   6.2913e-01 -3.1084e+00 -1.7450e+00 -1.7269e+00  6.6084e+00 -1.3424e+00
#>  -3.9579e+00 -4.4254e-01  9.5913e+00  3.7979e+00  1.2263e+01 -1.1058e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]