Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.1123  10.3223   7.1291  -7.6570  23.7089   4.7698   7.8746  -1.7714
#>   -1.5087   5.8759   7.5852  -4.5684 -13.0411   5.7218  -8.1918  -9.4140
#>    3.2437   5.8782  -5.3063  -4.0056   9.7451  -5.6930  19.6997   1.9400
#>   -4.0102  -0.8613   1.0838  12.0242   8.3772  -5.6623  14.6359  -1.6431
#>   -0.2404  -6.4609  -1.6947  21.6150  -2.3571  -1.8509   1.8057  15.4373
#>    2.5480  -8.0586  -0.2900   9.4225  -4.3885   4.4902   2.9541   1.2626
#>   -6.2076  -4.4644  -6.6543  -2.4974  -3.7161  -5.9221 -13.3543  -4.4236
#>    0.1552   2.3722  -1.1919  -8.9971  12.7462  -3.4853   6.5427  -6.9842
#>    8.2244   1.4793   5.3169  13.6816   7.7185   2.2089   0.9000   2.9626
#>    3.1858   3.3670  11.5887   5.8841  -2.7770   3.8679 -14.3024 -11.2376
#>    7.9266  -8.0394   2.5070  19.5012   1.8493   8.0673   4.0832   9.4396
#>    3.9544   5.4735  -1.8821  -7.1288  -1.0820 -10.8934   4.4672  13.1752
#>    0.4779   6.8313  -6.2069  -1.4653   4.8074  -7.5090  -8.1177  10.6981
#>    6.1922   6.7506  -0.4804   8.3078  -5.9987  -7.1506  12.4259 -19.9817
#>    7.1517  -6.2217 -16.3474  -4.0752  -9.2143 -12.1985  -2.2102   3.0767
#>    8.2691  12.8916  -6.2538  -4.6375  -2.9614  -7.4122  12.3654  10.1219
#>    3.7768   4.1357  -8.5593   7.1802 -13.8414   2.6256 -12.7341   7.8952
#>   -1.4472   2.2021  11.1648  -3.6374  10.2806  -9.3944  -5.3361   0.8998
#>   -4.4557   6.4142  -0.5078 -14.8373  -0.6310  10.7559  -6.3624  -0.6293
#>   -2.1012  -1.7291  12.4486  -4.3502  -2.2536  -3.7423   0.5782   8.4543
#>    0.6189   4.3418  -3.9313  -1.6441   1.3030  16.7442   7.4747  21.9916
#>    2.1041  -3.2383  -0.5006  -3.4918  10.4644  13.5881   9.1917  -4.1111
#>   -0.2389  -4.7449  -3.6638  -1.0522  13.2165 -11.2009   5.4608 -16.1875
#>    6.3533  -0.5155  -0.5848   8.1884  13.9937   5.0340  -3.7154   9.4799
#>   -0.7065   1.9519  -3.2575  -9.3661  11.9121  -8.9749  14.1161   4.0528
#>   -2.0961   1.9953   2.2476  -2.8303   6.4295 -14.5320  -5.8624  -4.0446
#>   -1.5957  10.9838  -0.3776 -14.6038  -5.8301 -12.3786  -7.3307 -10.0620
#>    4.2632   4.1978  -0.1576   3.4119  19.4042  -4.2071   1.6358   8.9139
#>   -5.1980  -5.8871   8.6005  23.9278  -7.5160  20.1109 -20.4773 -12.9652
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]