Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.3464   6.0870  -1.2981  -0.1155  -3.9241   1.6422   2.6985  -6.1945
#>   -4.2953   8.0887  -1.5799  -4.0553  -2.8118  -3.7839 -11.0698  -9.4590
#>    0.1251   4.3141   5.8303   1.1582  -4.9950  -1.6537 -18.1200   3.3755
#>    4.7144  -2.9440   2.5842   1.2160  -9.8603   7.7279  12.9965  -0.5479
#>   -3.7720  -0.6213   7.9917   2.1371  -7.4731  -2.5648  -2.8200  -1.3162
#>   -0.4730   3.3049  -5.9082  -2.5553   6.1649  -6.5306 -17.8235   1.2225
#>   -1.1551   1.3136   2.4179   0.2098 -21.5705   2.6620  15.5292  -3.8190
#>   -2.4230  -1.1027   7.3448  -4.8937   1.5769  16.9229 -16.5170 -13.7631
#>    4.0922  -3.6073   3.7891   0.1544  -1.7347  10.8576  13.4918 -12.2711
#>    3.1280  -9.8418   3.3819  -7.8852  -4.1158 -11.4868  13.5976  -0.1252
#>   -0.9327   5.6807  -8.3789   1.0145   8.8177  -6.0081 -10.2654  13.5717
#>   -0.8454  -1.8161  -0.2026  -7.2593  -0.8863   4.1843 -10.7143  13.4604
#>   -0.9372   7.0915  -8.0246  10.9951   2.2991 -20.9933   0.4366   4.4461
#>    4.7687  -4.1252 -13.2647   5.4904  13.1201   3.0620  10.2846   1.3499
#>    1.8315  -5.4108   3.0579 -12.2176  -6.7516  13.9838  -7.3905 -17.9585
#>   -0.3527   5.3803   3.3794   2.8351  11.5348  -4.2135  -9.5280  18.3461
#>   -3.0269   1.7763  -7.5605   6.4257  -2.8510   1.3778   9.3037 -10.8700
#>   -1.0131   3.9477   5.0145  -8.9042 -19.0391   9.5839  14.0894  12.9809
#>    2.4335 -10.0491   2.3240  -9.6233   0.7500   4.5414   4.5425   6.5815
#>    0.4831   1.1362   8.9360  -0.6620  -4.3886 -17.1635   9.7602   1.3619
#>    0.8976   5.4767  -6.4405   0.7489 -18.3691  13.8035  -0.2847   3.8951
#>    1.5458 -11.7048   5.5278  21.9545   5.1320   0.9106  12.4685  -5.4318
#>   -2.2499   8.7910   9.7140   2.0885 -21.7765   8.5345   0.3731 -10.3457
#>    1.3224   2.2987   0.9159  -4.5440   3.3215  19.0971  -6.7341  -3.1431
#>   -0.2723   6.5527   0.6107   3.1092   3.2116   3.5159  15.4783   1.3744
#>   -3.4680  -1.4444   6.0580  -1.0545   2.3662   6.8509   4.9352  -2.7342
#>   -1.2853   6.5169  -0.3054   6.1274   7.7247   2.3044  12.7011  19.7399
#>   -1.1431  -3.1095   0.4711   9.2223 -11.9684  -9.1603   1.5865  22.8390
#>   -2.3108  -1.6859  -6.7666   7.9604  -4.6527 -12.6880  -1.7029 -11.1740
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]