Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.2662   1.0681   0.2243   1.7029  -3.4790  -3.1632   3.0101   6.6573
#>   -3.0410  -4.4874   6.3710  -1.0760  -1.9630 -11.3641  11.4776  -1.5828
#>   -4.1383   6.7229   5.3159   3.8973   1.2793  -7.5721 -14.6363   1.3636
#>   -4.0742   4.0005  -4.7143   2.8475  15.5699 -12.0845   0.9687  -6.8828
#>   -4.2279  -0.0514   6.0345  -3.5179   3.0945  -3.6582 -20.8695   6.5030
#>   -0.3559  13.2781  -8.6712   6.2018  -5.5314 -11.0847  15.3060  -3.6255
#>    1.1695  -3.8705   7.1148  -4.0189  11.3312  11.4210  10.8631 -14.6166
#>    1.4882  -3.6738   2.4306   4.1396  -7.7060   4.9382 -18.9789  10.6090
#>    7.6172  -4.4516   1.8996   3.2535  -1.9973  -8.1795   0.3374  -0.2613
#>   -0.4145   0.0625   2.8247   2.2171   5.2954  -6.0057  -3.4361  16.0021
#>   -1.5395   4.5639   5.5485   8.0341  -5.0156  -4.1117 -16.1642  -0.1206
#>    2.4146   3.6285   0.2371   6.5931  -3.1873  -5.9985   5.2931  -7.5902
#>   -4.6512   1.2915   1.7054  19.1238 -12.9796  -0.8501  -7.1200   0.9311
#>   -7.2376  -7.7405   6.9246 -14.3299   9.1012  -1.1897  12.0277  -4.2425
#>   -0.5070  -3.0598  -8.8683  -4.6678  -4.8653  -4.4190  17.3720   3.2815
#>    2.7318  -8.8038  -0.0070 -18.3424  -5.7666 -11.2627   1.4710  13.7093
#>    3.5783   3.2168  -5.3175  -1.3902  14.3023   4.2572  -0.5644   3.9841
#>    2.8074  -1.7070   0.3483  -1.3639   1.2426   3.2344  -9.2657  -3.4040
#>    1.5171  -2.6883   6.4204  -6.4356   1.6663   7.5400   7.0943  -9.6731
#>    6.5585  -3.0890   3.1052   9.1984  -4.6195  -4.9674   6.4438  -3.0530
#>    4.7957   5.4658  -0.0648  -8.8272  18.2268  -6.4999  -2.2720   2.1036
#>   -2.8529   2.2893  -1.6891  -3.7953 -18.7788  -5.0562   0.7735  -7.2435
#>   -4.7276   9.2307  -5.2660  -4.6204  -3.4807  14.7610   7.9375   2.1426
#>   -6.0615  -5.1421  -8.5525  11.3841  -6.4726  12.7147 -10.1403  -6.6132
#>    0.7730  -0.9192   5.3797  -7.7169  10.8469 -22.8725  12.2137 -15.3078
#>    1.6023  -3.7958   5.8465  -7.9332  -4.5290  -2.1191  -8.2947  11.2329
#>    0.9599  10.7647  -1.3403   6.0211 -15.3858   7.7669  -7.0645  -0.7383
#>    2.7137   3.9271  -1.7691   7.7947  -4.1504  -6.6203  13.4768   6.9631
#>    5.5362  -1.3099 -10.0038  10.3310 -10.4144  15.8725  -6.9801   9.3915
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]