Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.2093  -2.4862   8.7974   2.0700  -1.6681   0.4642 -16.4227  -6.5998
#>    0.4110   6.9719   1.7720   5.3647   8.9043  10.3497  -7.0010   8.8808
#>    1.5277   9.5479   0.0643  -4.5310   1.9779   6.7029 -12.4705   8.6810
#>   -2.4913   1.2101 -11.8886  -7.0070 -13.6864  11.6004  -4.1327   1.8065
#>   -1.2810   9.8556  -4.9378  -0.6958  -0.5698  -4.5757   5.6444  -2.7884
#>   -4.0749  -4.1276   2.4366   2.7863  -7.1051 -15.0907   1.3790  -4.5403
#>    0.7359  -0.9095   6.6085 -15.7966   4.9491  10.1774 -12.9978  -1.4261
#>    1.8047   3.7689   4.2096  -6.3528 -16.3751  14.2515   4.3479  -8.8352
#>   -0.2882  -0.4737   5.0938  -8.3096  -5.1352 -11.0653  -2.2699   4.7428
#>   -3.9388  -4.9426   4.0518  -9.4112  13.3874 -14.3671  -7.9657   7.9620
#>   -5.4310   8.8310   5.5486  -1.7466   4.5730  -2.2416 -13.1111  10.2693
#>    7.4046  -6.5183  -1.9235   7.3689   0.2728  -5.1581   6.9912 -11.5627
#>    1.7978  -0.2119  -1.4906  -0.4199   4.4095  10.8330  13.6181  -0.1525
#>   -0.5140 -10.0010   2.5992   4.4880   5.8043  -5.8332   6.8070   2.3162
#>    3.8159  -3.4106  -8.5306   4.9580   9.3659  -8.9881  -3.0352  -5.9013
#>   -0.8528  -5.6083  -2.5545  -0.4263   3.6598  -1.4513  10.1853  -9.0119
#>   -0.2466  -0.5111   0.3451   7.4366 -12.8587  -6.0232   4.1554  -3.0251
#>   -0.0607  -0.1692   0.2244  -6.3283  16.0720  -0.6862  -5.0458   6.0276
#>   -2.7807  -1.6713   0.6545  -4.3055  -0.5863  14.8439   0.4862  17.6501
#>    1.5325  -3.8564  -9.7258  13.0599  -5.5651   8.1872   1.1583   3.7496
#>   -1.0013  -2.2497  -2.9326   5.7763  15.5983   1.6812  13.5936 -13.7762
#>   -0.4498  -4.6383   1.6247  -1.3204  17.3931 -11.1882  -6.1137   3.9962
#>    0.2856   2.0239   2.8399  -8.1857  -1.8827   2.1727  -0.1947  15.9847
#>   -1.2970   1.9947  -0.5740   0.2444  -8.5748  -5.4688  -0.6096 -20.9731
#>    2.0309  -6.6341   2.8972   1.2909  -4.1576  -2.6774   1.0620  -4.3812
#>   -2.3802   2.1024   5.4041  -7.8652  -8.7037   9.1216  -9.8699   8.7857
#>    1.4649  -3.7282   3.3961   4.4939   4.2767   5.5049  -1.8201   6.4445
#>   -3.6264  -3.8301  -3.2825  -1.7809  -8.4603   0.5716  -2.1974 -17.8671
#>    4.3329  -7.8887   3.0968  -1.7133   3.7211   8.7596  -6.3385  11.1271
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]