Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.5772  -5.5781  13.9286  -8.7563 -12.5211  -6.4474  11.9700 -11.6974
#>    4.1720  -8.7961  -8.6715  11.5157 -12.1609   0.4712   5.2046  -5.7286
#>   -3.3526   3.9926  -0.9220   7.1069  14.4403   4.3348   4.8442 -13.3355
#>   -3.0464  -3.4139   7.4905  -1.4734 -12.0496   7.7754  -6.8280   3.8843
#>   -3.6596  -1.6362 -15.3930  -1.4624  -4.3322   3.2745  -8.0553  -1.2927
#>   -8.3573   3.2124  -4.8474  -4.4074   3.6079   7.0024   2.5787 -12.5006
#>    2.5288 -13.3254   1.6152   0.6870  -0.1844   5.6814  13.5012  10.8459
#>   -1.6750 -10.0065  -7.2446  -5.1899  -4.7831  -5.7551 -14.6143  -3.0667
#>    0.0473  13.5807  -0.3229  11.3154  -3.1953  -4.5954 -12.8313  11.2477
#>   -2.0664   2.5820  10.7919   7.2920   6.8399   9.0990  10.7827  -8.1934
#>    5.0256   0.4616   9.8296 -13.9047   7.2660   4.2158  13.6424 -11.5989
#>   -1.0985   4.5203  13.3896   3.8718 -12.0228   3.5099   4.9435 -10.0901
#>    3.9564   5.9240  15.8620   2.9067   4.0814   1.8900 -18.2691  -9.0812
#>    2.1615  -6.5207  -4.0391 -13.7825   1.5580   3.6787  -5.6470  -8.1597
#>    2.5624  -3.7219   9.5529  -4.1807  24.9107   1.2387  11.7809   1.1635
#>   -4.4468   4.5968 -16.7241  -3.5423  -5.5058  13.3726 -19.8569   1.5944
#>   -1.4644  -1.2882  -7.6829  13.0768  14.2868  12.8236  -1.3153   0.1129
#>   -0.4256  -4.9597   2.4788   8.1898  -5.4138  -3.5485  -2.8557  -2.5309
#>   -1.0462   6.6641  10.7774  16.4219  11.3908  -2.6795  -1.1284   6.7307
#>   -0.7813 -12.6532  13.5003  -0.9381  -4.4019   1.4157  11.1079   4.1063
#>   -1.9447   5.3318  10.1078  -7.5710 -17.0823 -10.6538  -6.1707   1.4408
#>   -1.0014  -7.2940   9.2701  -2.7105   6.1921 -11.2566  -8.5932   0.7707
#>    0.8527   3.2914  13.8437  -1.8666  -4.0924   0.4016  -2.5709  18.1343
#>    1.5141   2.4585  14.6316 -11.0644   9.1395 -24.4135   2.5907  -1.3358
#>    4.7977   4.0708   1.0334  15.9684 -12.3820  -1.2303   4.8662   4.4639
#>   -0.4398 -10.3394  -5.5940 -34.3490   2.9946  -6.5880  -8.9939  -1.7254
#>    6.0428  -8.2555  -2.5503   8.5340   8.5366  -9.7839  -6.6943  -0.8406
#>   -4.5411  -3.8270  -1.1455 -12.9979   0.4122   9.2904 -18.1835   3.9915
#>   -2.3513  -7.7742  -8.7226  14.4664  -9.9273   1.4207 -10.0724   3.1637
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]