Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.4539 -10.0176  -5.4331 -11.6924  -1.2154   1.2281  -5.6785   1.0011
#>    1.6298   0.4993   1.7981  -1.1188  -3.9787  -8.7272  -6.8364  -4.1657
#>   -1.4250   4.5296  -5.1198  -4.7513  -4.8829  13.1111 -20.7039  -4.4994
#>    1.9825  -7.1703   8.1587   2.2895  -0.7767  -9.6995  -5.6600   8.9178
#>   -3.1990   0.9128  -3.6880  -3.1639  -5.4215  -5.7170   2.6888   7.3965
#>    6.7760   3.3317   5.5459   3.9056   7.6686   7.3095  -8.5897  -0.1554
#>   -0.6986   5.0317 -10.4517   4.9839  -8.8424   6.6078  11.0523   2.3773
#>   -4.2698  -1.5502  -0.3122  11.2828  -1.8401   4.5911   1.1048  -3.5439
#>    2.7383   3.5119  -1.3815   8.2268 -11.3025  -2.8000   4.1185  -4.0563
#>   -5.2401   2.4650 -10.7157  20.5853 -10.5417   2.1354   4.5710  -2.6819
#>    2.3889   4.4043   4.6083  -0.6399  -1.4698   7.0934  -5.4311   6.5308
#>   -2.7324  -5.9957  13.9890  -1.4469   6.5626  -4.9785 -14.6368 -15.8764
#>    0.3790   0.6710  -0.2362  -3.4655  -2.3222   4.9041   3.1936 -10.9933
#>    2.3955  -2.6907  -6.4238   4.5231   5.4835  -0.6045 -10.8883  -2.7163
#>    1.5899   0.0547 -13.4763 -13.7296   0.3751   2.7203  -4.0666 -18.9992
#>    0.5298  -3.0922   6.8645  -0.7940   5.2168  -8.3244   7.3022  13.2022
#>    1.8895   4.1119   1.1615   4.6497   9.5399  14.2144 -16.6177   2.7079
#>    0.1745  -0.8587  12.3310 -15.1861  14.3971  -9.0931   3.9944  -4.2908
#>    1.3152   2.3897  -2.8863   1.9695  -4.1414  -8.8739   3.6375  -6.4035
#>    0.1896   2.4877  -5.6915  -6.9597   6.3127  -3.0399  -1.8937 -12.7286
#>    4.8388   2.4439  -1.6660   7.1711   1.5728  -9.5513 -15.9859  -5.7709
#>   -1.5930   2.0217  -9.1697  -1.8796   2.4859   6.8223  -6.6329   1.3861
#>    1.1176   4.3879  -0.3508  13.6838   2.2225  -6.7058   3.5721   8.7161
#>    0.8222  -3.8556  -2.9712  -8.5060  -3.2681  10.4799  18.9686 -11.0568
#>   -1.2700   1.6297  16.6349   3.2625  -4.5583  -0.2375  -5.2658  -0.4159
#>    0.3623  -1.4269   0.8156  -8.7879   1.2743   6.9719  13.2746   9.1988
#>   -2.4120   4.6997  11.9487  13.7271   4.4021  -4.2456   0.8970  -5.7276
#>   -0.2639  -2.1008  10.8920  -9.5338  -2.4717  -0.2050  -4.6730 -13.4698
#>    2.4957 -13.0685   4.5010   0.4700  -3.6921 -11.0222  -5.1889   4.0273
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]