Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -2.8986e-01  6.9775e+00  4.2943e+00 -1.1458e+01  1.3911e+01 -8.6037e+00
#>  -4.9723e+00  2.0899e+00  1.7806e+01 -2.1201e+00 -8.8977e-01 -6.2331e+00
#>   2.8133e+00 -7.7398e+00  5.3749e+00 -8.6602e+00 -9.2835e-02 -3.1893e+00
#>   4.1755e+00  3.5634e+00 -2.5531e+00 -1.2456e+00 -4.4903e-01 -1.0641e+01
#>  -5.3320e+00  2.6210e-01 -1.8126e+00  6.1045e-01  5.5388e+00  1.6582e-01
#>  -5.2470e+00  3.2720e-01 -3.0657e+00  2.1905e+00  1.6892e+01  1.1414e+01
#>  -3.3340e+00  1.2819e+00 -5.7057e+00  5.5320e+00 -8.5452e+00 -2.6770e+00
#>   4.8194e+00 -1.9410e+01  1.0230e+00 -4.8053e+00 -3.0577e+00  1.5556e+00
#>   8.2430e-01  1.5253e+00  2.1755e+00  1.0253e+01  1.0435e+01  1.1607e+00
#>  -8.7752e-01  3.2110e+00  1.1430e+01  1.4049e+00  2.1350e+01  9.7065e-01
#>   8.2544e-01 -2.4499e+00  6.3446e-02 -5.6639e+00 -1.1328e+01  6.2045e+00
#>  -5.6884e+00  6.3861e+00 -1.9351e+01  1.1279e+01  3.9130e+00 -1.4433e+00
#>  -5.2376e+00  1.0892e+01 -3.6314e+00  9.6831e-01 -6.5040e+00  5.3155e+00
#>  -8.3475e+00  1.0797e+01  3.0997e+00  2.7899e+00  4.4091e+00  9.8556e+00
#>   3.4251e+00 -3.7549e+00  1.1867e+00  6.0014e+00 -8.6927e-01 -4.7571e-01
#>  -3.4974e+00  1.4256e+00  3.3350e+00 -1.3479e+01 -3.2887e+00 -1.6029e+01
#>  -7.9821e+00 -3.5470e+00  1.0398e+00 -2.5508e+01  7.4568e+00  8.9943e+00
#>   4.7799e+00  5.7861e+00 -1.6681e+01 -2.4428e+00 -1.1039e+01 -5.4679e+00
#>  -3.1168e+00  3.5725e+00 -9.8606e+00  1.8081e+01 -2.8589e-01  3.4068e+00
#>   3.4457e+00 -8.3406e+00  7.1117e-01 -8.2863e+00  9.1601e+00 -4.2287e+00
#>   6.6559e+00  1.4111e+01  2.0201e-01  4.5929e+00  2.6990e+00  2.4731e+00
#>  -3.2653e+00 -4.1153e+00  3.0082e+00  8.3747e-01 -1.1879e+01  5.5041e+00
#>  -1.2152e+00  1.4537e+00 -5.0834e+00 -1.1682e+01 -1.1811e+01 -1.0861e+01
#>  -1.5443e+00 -5.3810e+00 -4.1494e+00 -5.3425e+00 -2.1714e+00 -4.0140e+00
#>  -5.6840e+00  1.9383e+00  4.8899e+00  8.1429e+00  6.9667e-01 -3.8095e-01
#>   3.8226e+00 -4.2020e+00  2.6419e+01 -3.7582e+00 -2.5187e+00  1.5439e+01
#>   1.2016e+01 -1.5928e+01  1.1408e+01 -1.6755e+00 -1.4279e+01 -9.6002e+00
#>   5.5968e+00 -4.6213e+00  9.8742e+00 -3.9905e-01 -7.2905e+00  1.9517e+01
#>   6.6718e+00  7.2561e+00  1.8949e+01  1.4819e+01  1.0210e+01 -1.4827e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]