Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -6.5673  -0.3332  -0.5770  -4.2658 -19.8405 -10.6274  -8.1995  -6.0153
#>   -8.3991   7.1042  -1.1185 -12.6647   3.6371   5.1394   6.1877  -6.7740
#>    1.0222  -1.4823   4.3661  -5.9264  -4.7064   8.4295  -0.6827  -2.3471
#>   -6.1752   1.0268   7.3046 -11.2438   3.3122  -3.8867 -16.7491  -2.5357
#>   -1.4756  -5.1460  -3.9841 -11.6285  -2.7846  -8.8416   3.0254   1.6382
#>   -5.0527   2.2413   4.4064   0.4533  -8.4799   4.7682   0.0492  15.2338
#>  -10.5434   3.0147  -4.9144 -14.0743  -7.2347  12.1225  -0.4296  -6.9605
#>    3.6998   2.5143   6.3792  -1.4839 -13.3012  -5.2860  -6.8993   0.8990
#>    3.1930 -11.2391  -7.0882  18.3860  14.3370   2.6598   0.8095 -16.3738
#>   -2.7338   0.3787   2.6306  -2.2876  -3.5018 -15.6651   2.3424   0.2338
#>    2.8596  -1.0730   3.1532   1.4980   7.6113   4.7867   6.2289  -0.0641
#>    0.5004   4.3537  -4.4413  -0.3630  -0.8634   1.8528   8.4198   7.0853
#>   -5.9977  -3.9569  -1.4161   6.8340  -6.6574 -12.9651   1.4782  -6.7397
#>    8.3758  -0.1418   1.6552 -10.1113   7.6074   3.4850   4.3776  -8.0816
#>   -3.9619   4.8501   9.4946  -8.5341   0.9397   3.6639  13.9250  15.9248
#>   -3.7500   3.6927   0.4832  -2.1925   0.3130  -0.0146  -2.2890   9.5251
#>    4.5264  -4.3407 -14.6240  -0.5573  -4.4138  13.7921   1.5962   5.3934
#>   -3.0187   7.1674   9.6140   2.7427  -1.6003  -1.1231  -8.4988  -4.0927
#>   -2.4217   4.8715   0.8933   7.5329   5.9211   1.0919 -16.8882   5.0118
#>    8.5806   2.2497 -14.7194  -4.0174   5.6487  -0.7222  12.0750  -4.3927
#>   -1.0646   3.8167  12.1999   7.9480   3.2623  -0.6899  -1.4083  -5.0603
#>   -1.4325  12.8215  -8.9043 -13.4188   2.5359   5.2288   4.5644   3.3379
#>    9.8854  -3.8506  -5.9966   3.5007   9.7821   3.5260   7.7378  -7.3891
#>    2.5769   5.6926  -1.0350  -7.0029  -3.5436  -1.9453  14.7306  11.5310
#>    0.9074  -1.4557   0.7874   9.0528  -2.4450  -0.3743 -16.0893  -9.9578
#>    4.5748   1.5425  -5.4852  -4.3850  -6.1410   0.3156   1.6463  -5.3528
#>    1.9997  -1.1150  -0.6630   1.4570  -2.0362   0.8163  15.4826  26.9498
#>   -5.4321  -0.1744   2.3032   5.8185   8.9983   5.1771   0.9284   7.9018
#>   -3.1385  -0.4947  -7.4519  10.0449   4.3462  -5.3328  -9.2848   0.9478
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]