Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.7580 -11.8277 -12.7257 -20.4839  -5.7678  -1.3517   5.8571  -2.9234
#>    1.2719   4.5116   0.4260  -9.8002 -19.9570   5.6517   5.9966  -7.8935
#>   -0.2617   2.1676 -13.9853  -3.2029 -16.0984  -3.2916  -9.6673  -3.2246
#>    8.0954  -4.1645   1.3622  -0.3921  -8.5228  -1.0494  -4.8421   2.0676
#>    2.2314  -2.6971   2.2612  -5.4018  -2.1624  -4.8054 -10.5705 -16.4644
#>    5.6947   3.4147  -8.2093  -0.3771  -2.0714  -4.3755 -13.9936  -5.0478
#>   -3.7072   0.6239  -1.7126   1.4679  -6.7929 -11.6018 -11.4376  -6.1756
#>    3.2792   3.1971   4.7846   4.5354  10.6075  -1.7730   2.4317  -2.1676
#>    2.5863   4.4465   6.9286   9.2528  -3.8285  -0.5172   5.6410   0.1284
#>   -7.0236 -11.4360   4.2098 -10.1398   1.4962  -8.5233  -1.7586  -5.4950
#>   -8.0043  -2.8332   0.5233 -14.4967  16.4762  -5.0637  -2.4737 -10.1617
#>    4.1883   6.1411   5.3523 -13.7601  -1.2789  -1.1641  -3.3922  -1.3514
#>    5.5742  -9.2603  -8.1115  -3.3501  17.2961   4.2847   9.7226 -12.9441
#>  -10.4631  10.6339   4.7614  -6.6161   8.3123  -4.3121   7.0943 -12.2488
#>   -9.6796   2.0921  -7.7371   4.4980  12.7780  15.7059  -1.0830   1.9283
#>   -7.0924   0.5471   3.5126   7.2620   4.3361   7.1237  -1.0815   3.2556
#>    3.5874  -7.6849   9.8350   4.1420 -12.2853  -5.9134  -7.2864  -2.5173
#>    5.9088  -0.7827   8.8719   0.5104 -10.2698 -11.2998  -7.2858  -5.5704
#>   -1.2148   2.6152   8.8311  -8.7646   5.9274  -5.1664  -3.4789 -18.0988
#>   -1.2810   2.3972   2.9126  -3.7020   3.4267   0.6293  11.7554   2.4652
#>    4.9488   3.6387  -6.9079  -2.8425   8.7562  15.5018   4.7327 -10.3176
#>   -3.5914  -8.4557  -4.8023   7.9984  -7.6618  -2.0701   3.9443 -10.3344
#>    2.8836   8.5381   3.9074 -12.7475 -10.1950   1.6109   8.5469  -3.7514
#>   -5.4699   8.1038   3.9662   0.1592  -6.6925  -1.5526 -17.8106  -7.2758
#>   -4.5906   8.6434  -2.9072  -7.8120  -8.5212  -5.8502  -3.8912   8.4502
#>   -3.1183  -1.9210   3.7977  -5.7109   0.0663  -8.4214  26.8045 -10.8430
#>   -2.8479   6.9780  -6.6278   6.2273  -3.7190   6.5024 -10.9217  15.8143
#>    4.7175  -6.4825   8.8498  10.0394  10.4972  11.2632  -4.3131 -15.1820
#>   -1.3779   3.1893   1.5176   5.1614  -5.4378  -4.2737 -17.6776  -2.5180
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]