Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  12.3446  -0.2622  -1.2229  -5.8522   9.4910 -10.6253   4.4204  -0.9421
#>    8.4468  -1.7901  -2.2639   5.4105   0.0604  -0.6257   1.6270   7.8860
#>    7.3703   1.0041   9.8408  -6.2394 -23.2222  10.9443  -5.7633   8.9407
#>    0.5469   0.6831  -5.2188  -6.3724  -3.7950  -7.1392   5.4261 -10.0377
#>   -1.8841  -8.6797  10.7470   8.4017   2.7484  -5.1704  -0.1243  -2.3799
#>    0.9331   8.4650  -3.5766  -4.9009 -20.9297   8.3750  -6.9644   7.3181
#>   -1.2291   4.0904  -1.3685  -1.3620  -8.2534   3.2919  -2.8111  13.3702
#>   -3.0247  -4.7228   7.4313   0.3785  -3.5243 -15.1273   8.4868  -7.5430
#>    4.1146  -3.5235  -2.6317   8.7889   9.3614  -8.0614   0.3191   5.9265
#>   -9.2006   4.6700   1.7219  -5.5358  -2.2236  -5.2247 -17.4869 -14.8291
#>    2.2059  -2.4266  -2.4972   6.8476  -2.5504   2.2103  -2.3311  -9.3663
#>   -1.6019   2.0076  -3.3910  10.0407   0.5272  -0.8732   3.4619  -7.5628
#>   -0.7194   4.0903  -9.6092  -7.8201   7.8433  -2.2409  11.2322  -9.6487
#>    3.8331   2.9706  -6.2284  -3.0993  13.0326  -6.3968  -1.6145  -3.0233
#>   -2.1189   6.9643   8.6426   0.8317  -9.5076   8.4827  -2.0799  11.3573
#>    4.1859   7.9948 -21.3465  11.4821   6.3849   9.5691   4.3513  11.2859
#>    2.4423   1.3953 -14.5250   4.7830  -3.5880  -6.4280   0.1354   9.2090
#>    2.9995  -9.1861  -8.1589  12.3429   6.9700  -1.5016   8.9163   7.9592
#>    7.1045   3.7023   3.2080  -7.6577   3.7841  -8.4699 -21.8087  -7.9044
#>   -1.9509   0.6851 -11.3859  21.6950   3.3907 -15.9824  -3.2027  -2.2037
#>    1.4556   3.5187  -0.2440   0.2336  -1.1856  -0.7042  -5.0000   1.0259
#>    0.4786   3.3387  -0.8456   8.3683   6.4329 -10.1338   2.7871   7.4099
#>    1.0083   3.3899   4.3164  -2.6431  -2.9762  11.2438   2.8293   4.8815
#>   -5.8094   4.7293   6.6206  -4.3548   2.3133   2.5685  19.4075 -11.3136
#>   11.8313 -10.0730  -7.0546 -11.4151  11.1228  -8.0611  -3.1532  -2.5406
#>   -1.7708   0.5874  -7.5017   0.3842   1.6348   2.9361   0.9924   7.5723
#>   -1.6219   0.4127  -6.5192   4.4795  -1.0379  14.0236   4.8212   7.3675
#>    4.7944   2.9809  -6.3558   3.6985  16.1882   4.1084   6.8179   4.4104
#>    0.3469  -1.3371  -0.2029  -1.5409  -4.7523   4.7691  -0.9307 -13.8863
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]