Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  11.5585  -9.2223   9.6350   5.5964   4.3503  11.2423   4.1245  -4.0839
#>   -6.2799  11.3274   3.7410  -7.4196  -4.0601  -2.8330 -16.1587   6.4234
#>   -1.9207 -12.3280  11.1536 -13.6667  14.1073   0.5764  -6.2085 -20.0313
#>   -3.9961  -1.7572 -10.3654  -1.5056 -11.9862  -8.4665 -12.5519   1.8453
#>   -1.5039  -2.5501   3.6309  -6.4955  -0.1736   1.1437   6.4445  10.2443
#>   -0.3368   0.0593   4.4983   8.8346   8.0184   4.9612  -6.1940   0.2141
#>   -1.1794   1.2450   7.5007 -14.1501  -6.4305   4.8672   5.4926  15.2786
#>    0.9051   1.1859   7.7582  -4.6070  -0.2555 -11.3072  -2.0117  14.9040
#>    3.4651   0.4814  13.0568   0.4160  -7.5984  13.9574   3.2055  -8.8462
#>   -2.8642   9.0324   5.4462  -2.7645   5.8341  15.2735  -5.6700  -0.1374
#>    0.7700   7.6089  -6.9225   1.8924  12.6552  -2.9047   5.9066  -1.9064
#>    0.7429  -4.0615   2.5306 -15.3034   1.8901   8.1892   3.6066   9.2945
#>    0.4788   1.7095   0.1964  -0.8379   2.9043   7.6073   5.5846   5.1010
#>    1.1733   3.6127  13.4010   1.6063   4.7938  -5.6029  -0.9872   6.4875
#>   -1.1866  -5.5054   0.2801   4.1673  -9.6522  -2.4710  -4.8658  -3.4499
#>   -0.7665  -1.0788   1.5141 -21.0261  -3.3082  -1.5760  -1.9520   6.4465
#>   -6.6999  -4.6310   1.1531   9.2694  -1.7398   0.4601 -16.3949  -5.3009
#>   -9.1835   2.2795  -3.6015  -1.9283   5.2875  -6.8613  -2.4529 -17.3747
#>    2.7190  -2.3569  -4.1774  -0.1555  -9.4705  -9.1808   2.9917  -1.0229
#>   -4.3163  -4.1902  -3.7224  -3.7601  11.3692  10.6116  -0.5100   2.4159
#>   -2.2519   2.4905  -4.1772   1.9594   6.8355  -5.5989  -3.8881  -6.9298
#>   -4.5925  -0.5289  -2.8658   1.5719  -9.0250  -7.9758  -1.9099   8.0099
#>    1.7482   1.7982   3.0666   4.5727  -3.7998  -5.3002 -12.0319   2.4629
#>   -0.0432   1.0657  -1.6632  -6.5301   4.8339 -11.6588  -5.9360   8.5492
#>   -3.6597   7.5033  -2.3660  -3.5018  13.4197  -7.1038   2.4432   8.4447
#>   -0.5039   0.9857   2.5018  -5.8154  11.0813  -9.5667   0.0387   1.2399
#>    0.4888  -2.6480   3.0022   3.0679 -16.9203 -16.4418 -12.7271   0.9810
#>    3.4438  -5.0507 -11.0839  -1.4798  -0.2926  -9.4604  -2.6099   1.7085
#>   -1.9639  10.7478   2.8682   6.9850   6.5333   2.5747  -7.9535  13.5123
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]