Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.7074  -1.7831  -3.5322  -4.6565   3.6932   2.7043   2.4405  -1.2660
#>   -8.6508  15.5205   7.7555  11.1632  16.7900   3.3854  -1.0706   1.6639
#>   -2.0006  -7.2567  -5.6005 -10.4279  -0.0711  -6.1860  17.5578  -4.0756
#>    3.0948   3.2904   0.7287   2.8001  -2.2659   9.2413   6.8543   0.4438
#>    4.8776  -4.0639  -1.4025  -8.2406  -3.7565   4.2684  -5.7457   6.1635
#>   -4.2673  -1.7762   9.1574   0.8207  -6.4859   3.2451   7.4779  -6.9945
#>   -3.4265  -0.0206 -12.0675  -4.3442   2.0140 -11.3094   9.4181 -18.8487
#>    4.2760   2.6173 -12.6320 -12.7767  15.2467 -12.4737  -7.0849  -3.7037
#>    5.9327  -0.6977   0.5557   8.5471   1.8648  -1.4226  -1.4241 -12.8541
#>    1.3159  -4.4553 -14.8138  -7.0433  -9.5851  -6.4556  -4.8051  -0.7080
#>    2.6757  -3.5820   8.7476  -7.6002  12.1478  12.5926  17.4179   5.4941
#>    8.2046  -1.1581  -4.9981 -10.1554   7.2674 -10.5621  15.9863   4.6835
#>    7.2470  -8.7363  -6.7410   9.8187  18.3342   3.5387   0.2171   3.4004
#>    0.3539  11.2384 -13.8966 -14.8193   2.0532   2.9076   2.4456  -1.9817
#>   -4.1024  10.9638  -6.2797   3.8824   4.9510   9.7009   3.0021  -6.1348
#>   -6.9901   4.6571  -0.9951  -1.3335  -5.7568   3.6529 -10.2718 -12.7648
#>    2.2788  -7.0889  -5.0220 -10.2125  -7.2134  -1.1378 -12.1031  12.3647
#>    1.6639  -0.7286  10.0286   5.7376  13.7670   4.8750 -14.4812  14.5333
#>   -2.1585 -17.8592  -2.8187  -9.5139 -15.6296  -0.6457  12.0915   0.0735
#>    4.3406  -6.4358   6.1589  -5.9246 -17.4255 -11.6887  -5.0184  -8.8335
#>    0.8572   3.8588   5.9047  -8.0577 -14.2589  -8.3840  -8.0385  -3.0133
#>   -5.9180   8.4822  -2.5338 -12.6895  -1.6830   3.1292  -1.9631  -1.4497
#>    2.8348  -4.9429  -8.8588 -10.9522  -2.0351  12.6569   6.0132  -4.3544
#>    7.5558  -3.4788  -1.6886   0.8691  -4.5072 -14.9572 -12.7969  -8.2485
#>   -5.1233   2.9990  -1.9787 -12.3620   8.6837  -4.3157   7.7240   9.1382
#>   -2.9838   3.1465  -2.2979  -0.6926  -7.8688   3.5037   4.4826  -1.9703
#>    4.5851  -0.3620 -11.2591  -3.6313 -21.2275 -12.4381  -1.9033 -19.0031
#>    0.6543  -3.3416  -2.8401   2.1544  -0.9864   5.1139   7.9744 -16.0752
#>   -0.8770   5.6888   3.8065   5.4502  15.1561  -0.9020 -14.6103   1.7739
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]