Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.1776   3.5102   6.0791   4.7480 -20.4603   6.2893   3.4329  -4.0329
#>    9.1761   7.7666   5.0250  -0.0849   6.2616   1.3454 -10.4051  -0.1515
#>   -2.7679   2.4517  -7.6910   4.9029 -14.2497   5.5013 -10.3015  -2.6445
#>   -1.8012  -3.7724 -11.2788  20.4448  -8.0154   3.3364  -1.7571   6.3640
#>   -0.9204   0.3879   3.7870   0.1137  -5.8286  -9.4072 -21.2680   7.7897
#>    3.1772   4.9820  -3.8673   7.5521  -1.5206  10.3802   7.6955  -0.9120
#>    5.1068  -0.8439  -5.9491  -9.8826  -5.0488   4.2039  -0.2531   1.8144
#>   -2.5947  -1.3055  -5.7117   0.5609  -2.4053   1.1825 -21.3275   6.6103
#>   -7.9737  -0.7298   9.2359  -5.5525  -6.9016   2.4072  -6.6823  -2.8551
#>    3.4267  -0.0184  13.3597   1.4451  -4.3993 -15.3105  -0.8242   1.6445
#>    1.4711  -0.7875  -5.3446  -5.9351  -2.4048  -8.7845  -7.4782 -11.4446
#>    2.3658   0.0014  -7.1253   1.7668   0.6196  -5.2822  -3.6254  10.9554
#>   -0.0330  -0.9211  14.9434  10.0253 -22.0886   9.5257  -8.8870   9.2881
#>   -1.7676  -0.8560   1.3359   2.8351   8.9772   5.2308  15.2533  13.9404
#>   -0.5505   0.3585  -8.6812  -1.7477   6.8580 -10.3034 -15.3424   5.7826
#>    2.5205  -2.0453  16.1085  10.0113  -7.7893   3.4748  -6.1577  -9.2149
#>   -4.0607   7.9629  -1.2823   1.6446  12.0455 -21.0062   6.2462   6.3368
#>    7.3449  -5.3478  -8.8504   5.0386   8.4929   6.3625  -9.1954   3.7510
#>    1.1132  -4.1148  -4.1075   3.5612 -11.1880  16.1822   4.1426  -9.1758
#>    3.9034  11.0073   5.0057  11.8647  -4.1597   0.2330  -6.3126  -5.4387
#>   -0.3883   1.3270  -3.7698  -1.6286   7.6065  11.8949  -1.0684  -6.4608
#>    0.2432  -3.6293   4.9354   1.7205   1.5164   6.6759 -17.5407  -8.2520
#>    2.6021   5.2965   9.0518  -0.9373  -4.7545  15.7704   4.7333  17.2838
#>   -2.2496  10.1160   3.6882   2.1258   0.3463  -6.9280   9.3127   1.1702
#>   -3.0235  -1.4832  12.3679  -3.9702   6.4859  -5.8826  -2.7452 -15.2841
#>    7.8057  -6.3447  13.1125  -2.6176   4.8448  -1.9271  14.6724  -9.2333
#>   -2.5952  -6.9004  -5.7143  -0.6789  10.6783  -2.0170 -13.3790  11.1901
#>    0.1306   4.2188  -7.8479   0.7142 -13.3995  -3.1749  -5.1719  -4.5032
#>    5.0453  -3.6259   4.0104 -11.3925  -1.5282 -10.5384   1.8420  -1.5803
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]