Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.5735  -6.8278  14.0739   1.6330 -12.5140   1.7527  -8.3415   4.3600
#>    0.0951   2.2399   1.1862  -4.1728  -3.4025  12.3512 -15.5198  -7.7398
#>   -0.1230  -3.6286  -0.9280 -12.3845  -3.2143   3.2619   0.3428 -14.5769
#>    1.1746   6.9457   1.9741   2.0517   6.2592   7.5210   0.8748   6.0685
#>    1.5308   3.0807   2.5987 -13.9295  -6.6210  -5.4195  -5.0064  -1.7547
#>    0.3271  -8.2827   6.6349  -4.2082  10.0589  -1.8680  20.6313  -8.6420
#>    0.9718  -5.3418   7.0767 -10.8491   6.8183  -7.9923  -4.7744   2.5383
#>    7.9420   0.2814  -9.0070  12.8757  -1.9285   4.3300   1.0813  -4.8346
#>    0.6638  -6.8022  -0.6209   2.3805  17.9534   1.4676  -5.6245 -12.6668
#>   -2.1128  -3.5374  -3.2595   2.1472  -1.8466 -17.5074   0.1734  -1.9617
#>   -4.5074   5.8874  -5.3423  -1.9881 -10.1779  -3.3496   8.1586   5.2811
#>   -0.7215   1.9318  -1.5684  -2.3631  -4.6159  -0.0165  -6.0259   1.3685
#>   -0.0820  -2.4905  -5.2303  -2.1044  13.3029  23.7660   6.7805 -12.6022
#>   -2.9437  -0.6931   0.9519   6.5678  12.9279  15.1150  12.3452   9.2713
#>   -0.2430  -3.9271   4.5839   3.5582  -7.2895 -10.1946  -4.6109 -21.1126
#>    2.9347  -0.1520  -4.0583   0.2251  -7.7925   0.5749 -12.0635   0.9179
#>   -2.0287  -5.0406  -1.3876   2.6054   5.9249  -6.2182   5.9779 -15.5900
#>    1.4479  -3.5566  -0.7974  -1.7463  -5.8061   6.4724   6.2584  -2.1609
#>   -1.1315   0.3102  -4.2130   7.5689  -4.7728  -1.9907   1.2218   0.6905
#>   -0.8641  -3.7931   1.3542  -4.5769  -0.9394   1.1800  -2.3063  -7.4028
#>   -1.3581  -1.4704 -12.5721   1.4889  17.8381  13.5140  -0.8790  11.3651
#>   -2.9501  -1.9379   5.6043   3.4676  -2.5093  -0.4698 -14.6672 -17.3663
#>   -3.3306   4.9892  -1.2208  -0.7735  -4.7326  -6.1932  -0.4609   4.8866
#>   -3.3777  -1.8569   2.7230   0.3246   3.2209   3.4393   1.6984   8.5822
#>    1.0809  -7.4498  -9.8944   5.8413   0.5037   9.6907   7.0282  -4.1826
#>   -3.6247  -7.4156  -6.1020  10.5577   0.5019  20.4435   5.2578  16.3262
#>   -0.0784   0.4878   2.7205  -3.7065 -10.3050  -3.1459   2.7152 -13.3017
#>    2.7106  -3.9872   0.6432   2.7675  10.9602   7.9637  11.2025 -13.0465
#>    0.3632   1.0643   1.5989   4.9023  -6.7419 -10.4117   3.3004 -11.6596
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]