Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.3601  -1.9601  -1.4271   1.9676  -4.9426 -11.9319  -7.5462  -3.7408
#>    0.0135   3.4576  -0.7157  -4.8448   1.2482  15.6352  -1.1101 -16.3429
#>   10.3491  -2.0101   1.2200   0.6891   3.8657  -8.9415  14.6132 -16.7045
#>   -5.1451  -0.1566   2.7371 -12.8987   7.9263   4.2016 -13.4349 -15.9662
#>   -3.3640  -1.4859  -2.7196   3.4540  -3.9363  -6.0981   1.6133   7.4155
#>   -7.8362  -4.1921   1.4488  13.4360  -5.1162   8.4756   1.5721  -0.1358
#>    4.3068   0.2453   6.8482 -12.4563 -13.5476  -0.3468   7.7020 -13.1885
#>   -4.1753   3.7348  -4.1187  -3.9767 -12.6399  -7.2435   2.8693  -4.4684
#>    4.3269   2.6535 -12.0710  -9.7557   1.6697  -8.1143  10.8251  -4.4841
#>    0.4247  -4.7778   7.8465 -16.4419 -11.7220  -6.8675  -0.4821 -14.1173
#>    4.1448  -3.7204  -1.0268   2.7917  -5.7570  -2.6618   6.4332   6.4664
#>    9.3845  -2.9618  -4.6367  -0.7726 -13.2307   0.1829   3.7418   8.8582
#>   -3.8539  -3.4131  -1.6698   0.5096   6.4804 -12.9822   0.7341  14.8232
#>    3.8567  -7.6969  -0.8086  -7.6822  25.4896 -11.2315   6.4521  12.8693
#>    0.2470   4.5168   2.6686  -9.2340   4.9610  14.3494   2.6845   1.0474
#>   -4.8724   4.2078 -11.5835   0.6474 -13.2063   0.1179   3.4996   0.2355
#>    1.3088  -4.9709   2.1151   2.6680   8.8075  -8.6949  -2.6874  15.0999
#>    4.7153  -3.5180   6.9013   0.4132  15.7979  -7.7851   3.8283  -9.6894
#>    5.0372  -1.0086   1.8938   4.3459   7.4119 -11.6918   2.1007 -14.4356
#>   -5.0812   7.5472  -1.7231   0.2089  12.8021  -5.8093 -23.4257  -1.5031
#>    2.0673  -2.8241   2.5197  -0.5076   8.2765   8.8199  19.4079   7.4507
#>   -5.7577   2.8802  -1.4965 -11.3919 -20.6669 -11.2414 -10.1449  -7.5651
#>    0.4909  -1.7698   0.4259  -0.9152  -7.5733  -3.3228  -5.7792  -8.8862
#>    5.1589   9.9420  -0.0956   6.3549   2.5979  -5.7620  -1.1216   6.3966
#>   -0.7140   4.9980  -9.7745   3.9702   8.6420   9.9693   1.8523  15.3675
#>    9.9676   1.0632   5.9591  -8.2873  12.4393  -4.6692   3.8246   7.9735
#>   -3.1086  -2.5150  -8.5791  13.7973  -6.7346  -7.5590 -22.2379  -5.2326
#>   -3.9373  -6.8712  -0.2970  12.6912   3.8530   4.4579  12.4615   2.0238
#>    0.5388  -1.5626  -0.4538   2.2765   1.0612   5.2624   5.9425   9.0645
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]