Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.7880   5.1068 -16.5398   8.9020 -16.1910  -9.9925 -10.1570  10.0007
#>   -3.7545 -18.3641  -3.0617 -10.8555  -2.3310  -3.6727   6.0528   9.3160
#>    4.1199   4.0055   5.5934   2.0332  -1.4397  -2.6509  -6.1950  -6.7269
#>   -5.0360   7.0524  -2.4306 -13.0410   3.1356  -2.9288  13.1039   4.5983
#>   -3.0821  -0.1362   5.2662   2.3351   2.7124  -3.9781  -1.2791   8.1706
#>    1.0025   5.9474  -6.1638 -17.3274  -3.9738   0.8404   3.9865  12.7754
#>    3.2751   4.2219  -0.0784 -14.6016   0.5588  -5.7827   0.5871  23.5753
#>    5.1706   1.3551   1.6783   3.7012  -3.5084   4.0427 -12.6845   3.1134
#>   -4.3986  -3.7962   8.8610  -3.0471   6.1105   7.2568   8.1942  -3.5337
#>   -5.6752  -5.5781   1.5335   1.2892  17.6141  17.7334   1.7773 -15.3246
#>   -5.7746  11.2962   8.3627   4.0825   2.2612   5.9680 -11.4460 -11.7988
#>   -0.9649  -0.3017  -1.4486  -4.6321 -10.8296   3.0748  -5.5525  -8.5162
#>    4.0911   9.5696  21.8957  -5.1670   7.4064   2.6705  -2.2153   2.1183
#>   -3.8142   1.3226 -15.0792  10.5136   2.4827  -1.0847   2.6226 -10.5451
#>    1.1261   0.5232   4.3177  -9.9934  -2.2256   9.4212  -1.7187  -4.0208
#>   -2.4379  12.5999  14.2823  -5.1352 -13.5720  -5.2494  12.0325  14.6137
#>   -3.9094  -1.8565  -5.5407  10.6456  -3.1693 -13.6362  -8.7628   4.1282
#>    7.3402  15.5583   3.4408  14.1650  -5.0046 -19.1114   6.5629  12.3371
#>   10.5925 -11.5271  15.4254  13.4811  10.2073  -3.4548  -5.9493   3.3791
#>   -5.1815  10.1139  -4.5508 -12.2174  -5.6699   1.6528  11.7255   1.8632
#>    3.9921   0.4533 -10.6918   4.1865  -6.5682   0.1926  -4.1191  -4.3156
#>    4.9700   6.8785  11.1426  18.1344  -6.1303 -19.2041   0.5662  -2.5582
#>   -1.9050  -5.7052 -10.4364   1.0585   0.9993   7.4598   5.6203   6.2864
#>    2.8142  -0.4710   7.2722   1.7460  -1.2075   4.0987   1.9775  -2.9232
#>   10.3292  -0.9226   4.3273  -8.8005  -1.4868 -17.6709   3.5779   4.7525
#>    2.0476   4.3117   4.4353  -6.0814 -11.8545  -5.6453  -1.6325  12.3856
#>   -2.5769  -5.2248  -2.9123  -2.5597   2.7634  -3.1554  12.0614   0.1398
#>    2.0367  -5.6210  -1.5340  23.0499 -10.0569  22.4179  -8.8177 -11.6597
#>   -1.8769 -17.2117 -17.4126  -2.3379  -1.6475  -4.7256   7.0400  -4.9934
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]