Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.0261  -0.1145   0.4300 -13.7117  10.9917  16.3318  -3.0075 -13.1586
#>   -2.0399   2.5783  12.0969  -6.6391  -7.2488  -4.3593  -6.3794  -5.1238
#>    1.5754  -3.7870   7.2290  -3.6196  -2.7431  22.5864  -4.1593  11.9498
#>   -1.0245   1.5587   2.0463   8.4957   9.6422  -9.7781  -2.3755 -12.7567
#>   -0.7919   2.6534  12.8506  14.3506  -6.2904  11.2199 -10.3533  13.6712
#>   -3.8556   5.2934   5.7580 -11.5869  -6.4390   7.8319  10.1330 -13.4223
#>   -0.3374   4.0678  -2.4287   3.6438 -12.9200  -4.3832  22.2209   2.1681
#>    3.9645   3.3611 -14.0603  -2.2913 -19.5433   0.7319   5.7490   5.7127
#>   -0.5592   1.3937  15.6442 -10.2461   6.9215 -15.7184 -21.9285  -5.0594
#>    4.0732   4.2161  -8.5045 -12.8773   4.5981   1.1455  -7.5104  -0.7085
#>    5.9557   5.1633   2.5272  -0.7987   0.1414   1.4635 -10.7226  10.2261
#>    1.0350  -8.8884   0.3808   0.2485   3.3173  -7.8553   6.8730  -4.8649
#>   -1.2932   4.4195  19.5795   6.5424   0.4158  -4.0457   3.4145  -0.4736
#>   -1.8969   7.8536   7.1633  -0.5847  -5.9770  -4.2844 -20.9710   6.4615
#>    1.9200  -4.9956   9.5428  -5.9218  -2.9554   6.1284   4.8558   2.2996
#>    1.0080   4.7814  -1.6772   6.5901   3.2954  -1.6986 -14.0673 -21.9511
#>   -3.8845   1.4759  -2.2107   4.1632 -10.7515  -2.9086 -13.8133 -10.3421
#>   -1.9450   1.7463 -12.8296   0.6876   8.1125   7.0394   1.1551  10.0005
#>    4.3426   1.0159  -3.8476  11.0077   9.3120   8.3768  14.7469  -8.6919
#>    4.0575   0.3948  -0.6919   0.3474   4.4743  -4.4929   7.8527   9.1427
#>    0.9249  -3.7938   6.6057 -11.3351 -12.0057  16.1545   8.1769  15.5525
#>   -5.1447  -0.8197   4.5616  -7.5095   5.0590  11.9787   3.8216  -9.7748
#>   -2.1704  -0.2685  -0.2821   8.0634  14.5215  -0.0694   3.0712 -13.0015
#>    8.7572   4.9943   2.4546  19.9102   9.2313   0.5148   2.4013 -11.0543
#>    3.5905  -6.9610  10.2515  -8.9933 -14.3184  -9.4182 -11.1368  15.0567
#>   -1.4526  -0.9938 -12.9912  -9.0319   6.3803  -1.3222  12.4763   0.5116
#>   -6.7365   5.3855   7.6989  -8.6702  -5.7202   6.4144  -4.3044   2.8751
#>   -2.3865   4.4454  -7.8788  -5.6417   6.1971  -2.8348   2.3589   3.5846
#>    1.3139  -3.9369  -0.5461   3.9405  15.0720   1.6335  -0.6964   2.3501
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]