Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.4704   6.2224   9.0200  10.7479   3.6138 -23.8212 -11.6453  -0.9778
#>   -0.2542  -3.0947  -8.3259   0.5544   0.4600  -3.6559  -1.5297  -1.1412
#>    4.6041   1.7682  -2.6079 -11.3101   6.4722   0.7032  -9.5992   2.2047
#>    1.7341  -7.4838  -5.3745 -12.5067   1.9294 -11.6320  12.9904   2.5175
#>   -1.8563  -0.1273  -2.9779  -5.0413  14.1296   6.2153  -5.6836 -10.0798
#>   -0.2025  -3.5472  -9.1149   2.0457   7.0845   8.9048  -0.5152  15.3172
#>   -2.9778  -5.1541  -5.9550  -2.1256  14.1354  16.2202  -2.8574  -8.8830
#>   -3.3578  -9.5953  -9.1467   0.6715   0.7819  -8.1658  -1.5574   2.1509
#>   -3.2566  -7.7852   3.1583  -5.7579  -3.6414  -4.7959   6.5243   0.2503
#>   -2.6107  -1.6018  12.4075  -9.3800  -4.0206  12.1201 -13.1961 -16.9654
#>    2.2650   4.6177   1.2587  -2.4585  -4.3581   9.2865   6.1470   9.4684
#>   -0.6141   0.1104 -11.3272  -6.9679  -9.4684  -2.8398  -2.2277  15.7591
#>   -2.6637   3.4261   2.3643  -2.0490   3.0387   8.4539  -5.2383  13.5538
#>    2.2545   0.8350  -2.1843  -5.9093   6.3412  -8.3670  12.2682 -13.1416
#>    3.6630  -1.0609   1.9087  10.6348  11.1750  -4.3846   3.5103  -4.0355
#>    0.5590   0.4870  -9.9392  -1.2333  11.1044  18.4589  -8.8921  -1.1621
#>   -0.8621   3.0414  11.6450   4.2060  -6.2489   2.0322   1.3306 -14.3861
#>   -3.5675  -8.0215   2.4602   3.5176   4.5894 -10.2606   7.8444   0.0403
#>    1.0850   1.8519   4.0145  -6.1544   7.5397  -6.5047   2.4728  -0.1544
#>   -3.0434   3.7191  -1.1685   1.2782 -11.1582 -11.6536  -6.0839  -0.5065
#>    3.0057   8.7843   8.9611  -6.8201  -5.4652  -4.7503  -7.7098   1.9230
#>   -1.9530 -11.4159  -8.5003   5.5670   0.0674   4.0016   3.7780   2.3285
#>   -1.8089   1.3968   7.3674   0.5485   9.5092  -5.3372 -22.9955   4.0317
#>   -3.8780  12.2111   6.7546   4.7734  -4.1996   6.9871 -14.7990   5.6731
#>    4.5458   4.7825  -4.3399  17.5251   0.8412   1.1777  -6.2366   4.6244
#>    0.3262  -3.1711  -3.1719   7.4417  -0.5329  -3.0804  17.7069   7.8696
#>   -3.9782   3.0024   3.5562  15.0452  -0.3060   8.5724  -5.6607 -20.0199
#>    2.3610  -4.6243  -6.1567  -3.3783   0.2278  -4.4752   0.6645  16.2923
#>    0.8092  -2.6391  -2.7870   1.0505   6.8741  -7.1811  11.1484 -10.3331
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]