Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.7639  -4.5733  -0.0234  -9.7615 -17.9211  -2.4199   7.1507 -13.1471
#>    0.2969   0.1088  -0.4099  13.7597  -0.4451  -9.6295  -4.2226  11.8906
#>    1.7776  -2.0809  -0.5000  -1.7728  -0.4420  19.0845  11.2565 -21.5816
#>   -5.4670  -1.6331  -3.0865  -9.1035  -9.4813 -10.0846   2.3695  -2.7846
#>    4.0021  -3.4770  -8.3319  -7.5223   0.9242   1.1639   7.2788  15.4508
#>   -2.9770  -5.1752   0.2164   3.0687   6.9538  -2.5423  -8.1701  -9.1983
#>   -0.3518  -9.3482  -0.7117   1.6580  10.7597  -7.7048  -2.9886   0.5401
#>   -3.6818   1.8836 -10.8200 -16.6503  -6.7736   1.7003  -7.3275   3.5673
#>   -1.8264   1.9633  -0.9150 -11.9875 -11.0208  -4.0366   1.3939   5.8728
#>    4.3032  -2.5130   8.0777  10.9893  -0.7702  -7.3634  11.3766 -17.1714
#>   -2.6259  -8.7146  -2.5444  -2.3113  -5.2697  -8.2828  10.8879   3.1821
#>    0.6334   1.3974   8.2146  17.0209  -9.9000  -0.0622  -0.1958  12.3246
#>    3.9965  -3.3916  -8.8882   2.5764   4.9626   7.6539  -7.4690  14.2383
#>    1.2358   4.1195   2.9047   6.8865  -7.3691  -1.1958 -15.0808  -4.0409
#>   -1.5201   2.6963   9.8834   6.4857   8.5433 -11.0612 -16.5115   1.3428
#>    4.8795   5.9805  10.2352  11.6879   6.6770  -2.3359  -1.8343   7.5657
#>    4.5440  -2.6318 -13.4509   6.6616  10.1172  -6.0783  -4.0698   0.6873
#>    8.1272   8.2661  -0.4729   3.6528  -9.2380  -0.3778  11.9253  -2.5526
#>    1.8079   9.9996   9.7433   0.2631   2.3200   5.5526  -1.2570   8.0100
#>   -0.6411  -2.0885  -0.6639  -0.8367   3.9785 -12.1503   5.9994   0.8644
#>    2.3186  -5.8798  -7.0507  11.2061   8.4261  -0.8830 -14.3021  -1.3571
#>    2.7241   2.4864  -8.1048  14.0351  11.2537  -2.7016  -8.1634  -9.8496
#>   -2.3168  -4.1437   5.8320  13.3282  -2.7886  -5.3950  21.0936  -4.1918
#>   -1.2026   2.6085   3.6859  -5.1550   4.1891   1.7578  -4.8887   2.3933
#>    1.1935  -8.1842  -1.2344   7.7919  12.2128  -6.0705   1.0667  -3.6470
#>    0.0490   1.2803  -4.3977   1.6733   8.7025   4.7510   3.6750 -11.8807
#>    5.0237   6.9589   1.6946  -9.8920  -7.7848  -6.1326  -2.8842  -8.6294
#>   -0.2091  -6.2101   5.4908  -2.4178 -15.7632 -11.2123  -6.2587   0.3196
#>    1.0579   6.1242   1.5097   6.9116   2.8900   4.7861  -3.6207   2.8935
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]