Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.1074  -1.0123   3.6222  -3.3623  -3.7735   4.7130  -9.6747  16.7285
#>   -3.0202  -2.9666  11.0307  19.2247   3.4956   9.1833 -10.9324   7.3981
#>    0.1227  -3.0846   3.8816 -11.9063  13.0942 -19.0080   5.4506  -4.9802
#>    0.5831  11.6549  22.3661   0.9997  -6.9716  -6.3341   6.9161  -6.0401
#>    0.5392   5.2162  -5.8955  -4.2497  -3.3599 -10.4788 -15.5643   1.8895
#>    3.3716  -0.7590  -5.8758  -3.3545  -8.8930   6.8816  -1.5790  -1.7574
#>   -4.7553  -8.3015  -7.8600  13.4963  16.9511   3.6499   8.5648  -2.2003
#>   -5.2119  -3.1674  -2.6492  -3.2563  -1.3745  -5.2047   1.2160  -5.7367
#>    1.0077 -11.0355  -8.0901  -4.7075 -21.8371   7.3371   0.4448   5.7678
#>    1.9679   7.1681  14.2938   0.0913  -3.1118  -1.1840  -4.0113 -12.3356
#>    0.9822   4.3138  -1.8982  10.2083  -7.6855 -15.7684  17.1851 -13.2334
#>   -6.6478  -2.0108   2.8341  -1.1352   7.0651 -10.2687  11.2586  -4.8134
#>    1.6981   5.3890  -8.7224  -4.2347   6.4217  11.7205  -0.6967   2.1082
#>    7.9883   8.6080 -12.2654 -14.7606 -14.3002   1.2807   2.2026  -3.9135
#>    1.2390  -3.7077   4.3409   3.9699  -8.9525  -4.4432 -11.0183 -10.9695
#>   -7.4464   2.5770  -0.3380  -7.6715   8.4871   3.0890  -9.4550 -10.2371
#>    7.5530   7.8978 -11.8765   7.7049   0.2924   5.1451  -0.6578 -20.8686
#>   -0.6780  -2.0526  -2.0645   4.7807 -20.5244  -5.9628  21.0058 -11.1645
#>    2.7962  -3.3350 -22.6072   3.5164   1.3302   4.0600 -25.9506  19.3818
#>   -0.9885  -7.0469  -0.6642   6.9679   1.2415   9.4584   1.4742 -11.3788
#>   -6.5934   0.0619  -1.7949  -2.9682   3.8152   7.6923  24.4990  -3.5744
#>  -10.4125  -6.9664   1.1710 -12.8779  -4.4283   5.1777   2.3124   8.7531
#>    8.4519   2.5803   0.1758   1.1154 -23.5422   0.1497 -23.8574  40.1702
#>   -3.3549 -10.5009   9.0541  16.1856   7.3796   4.4530  -8.9017  -1.1728
#>    3.3018   9.9793  -2.7501  -2.6380  -5.0339   1.9815   3.8408 -13.9956
#>    1.9553   1.9186   1.8432 -17.8583  -2.1837   7.9229   1.9362  -4.0753
#>   -7.6059  -3.2682   9.9993  -2.6128   1.7444   3.5564  -7.1798   4.1101
#>   -3.5438 -14.4106  -3.6505   6.7032  16.4228   6.2349   1.0034   3.0393
#>   -5.0816  -4.7090   0.5831   1.1850  -2.8770   3.4562   2.0903   9.6392
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]