Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.0904  -1.7688  -9.2883   3.0279  -1.6405  -9.1705 -23.7396  21.8582
#>    0.9777  -6.4539  -2.3765   5.2712  -2.6758   5.7919  -2.0778  -5.2033
#>   -4.8302   1.3481  -0.7955  11.3646   0.2043  10.1880  18.5058  -2.2381
#>    3.0874  -3.3265  -5.4659 -13.8644  11.6816   3.7463   1.5410   9.8794
#>    2.0903  -0.4584   6.8383 -10.5339  -4.6497   7.2316  -7.2375   2.4577
#>    2.1946   2.5999 -10.9492   1.6853  21.9962  -5.5970 -18.7978  -9.3622
#>    0.8225  -0.4099   6.4600  -2.0324  -1.8582  -5.9861   7.1894  -0.0367
#>   -3.8095  -1.7309  -7.2773  -7.5857  12.6410  -6.2232  -1.7532   9.4477
#>   -0.3361   1.2429  -2.6001   1.7688   4.6831 -11.2504  10.9773  -8.8964
#>   -5.5305   8.0111   5.1523  -2.0108 -25.5411  -5.8502   0.9349 -10.2801
#>   -3.1748   6.2162   2.8202  -1.6140   6.1492   7.0923 -12.6456  18.7780
#>    1.2793   1.3287 -14.8471   1.8479  -2.6248  -0.7800   0.1812   3.6922
#>    0.8270  -8.1789   6.4383  -9.3985  -4.0216  11.4941   1.6027  11.3200
#>    0.1161  -1.2356   1.7124  -4.2594  13.5213   4.4507   5.2349  15.9891
#>   -4.2648   6.3608  -6.1677  15.4190   5.7305  -0.8643  -4.6373   1.1824
#>    3.8825   7.7677 -12.3278  -1.3224  12.0671  -7.7532 -30.2364  -0.4243
#>   -1.0225  14.7482   1.9532  10.6676  -0.7311  -3.8872  -5.2394  -6.6834
#>   -6.1696   8.2563   8.0813   8.0711 -16.3657   2.8001   7.1165 -17.9327
#>    1.1096   6.2707   0.2613  12.4437  -9.4037  -2.9459   3.9024  -8.4025
#>    5.1009  -7.5396 -13.6430   6.8291  -4.7202  -5.5395   2.0019   6.2215
#>   11.7866   8.9723   5.1671  -1.6154  10.9814 -11.9960   2.6796 -13.3260
#>   -2.1598   2.4223   3.4932   6.6683 -10.2236 -12.7499   5.2925   1.9773
#>   -0.5987  10.3007  -3.9943   7.5056 -12.8083   9.6770  -7.5159   3.3605
#>   -7.3397  -5.3673   0.5953  10.0070  -5.4314  14.3707  13.5100  -4.0661
#>    1.3447  -3.4908   3.8883  -3.8892  -7.7209   6.4253  -9.7803 -12.3387
#>    9.3829  -7.3165 -10.9855  -5.6311  -0.3115  14.4631   4.5868  10.7730
#>   -4.7453   3.0459   6.3850  -9.2696   1.9510   6.8815  -0.3365  -0.3438
#>    3.1557   0.0874   2.4885   7.2663   0.4699  14.4486   9.6396  -4.2636
#>    1.6074  -8.8634   3.3586  -3.0579 -12.9479   0.0761  -6.2693   8.2738
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]