Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.3800  -3.5771   8.0087   4.4508  -9.0870   8.5713   0.9515  -5.8240
#>   -0.1381  -5.1173   1.9313  -2.7255  17.2470  -3.9905  -2.1936  -0.6230
#>   -0.3769 -10.3909  -3.1242   6.4671  -6.3175  -9.1029   5.0873  18.0661
#>    1.7329  -7.7193   2.1294   3.1313 -18.0656  -8.9462  15.2096  -7.7030
#>   -3.0410  -1.0014  -3.4736  -9.1425   6.7691  10.6592  -6.6747  10.4365
#>    0.3966  -0.1114  -2.0899 -13.3478  -2.5729  10.3469   2.6991  -9.7204
#>    2.6474  -2.5385  -9.1587  19.1869  -1.3296 -16.6584  -3.3863  14.7665
#>    0.7726  -2.9425 -14.2379  -2.7258  -2.9563 -16.2669  -6.4003 -15.6382
#>    1.8168  -8.8928   1.6501  -0.1416  -2.1177 -31.8577  -1.1262  -3.2755
#>    2.7379  11.7141  -4.5583 -11.3359   1.8400   4.4155  -5.7885   2.9526
#>    2.5675  -5.2357 -15.2805   2.5470  11.4826  -6.1372   1.9156  -4.5692
#>   -8.9715   3.4179   0.1091   5.0629 -11.5932  -5.3863   8.9072  -3.9673
#>    2.0721   9.7668  -1.1288  -2.9823   4.4824   8.1583   4.7928 -22.5161
#>    1.3893  -3.1141  -2.7388  -5.6177   6.2427  -4.7915  -2.0603  -4.5156
#>   -2.9402   2.3342  -7.7595  -1.8376  -4.6085 -10.0538  -0.5338   5.0483
#>    1.0547  -1.0364  -4.0866   3.3182 -16.7478   0.8109   5.6558 -12.1643
#>   -2.5516  -1.0524   2.9015 -14.0321   0.3665   8.4494  -6.0283 -15.0041
#>   -1.2908 -10.0816   0.7514  -2.8521   2.6172  -1.6506   9.9320   9.0983
#>    1.0032  10.8095  -1.4626   1.9343 -13.3770   0.8378  -0.3299  -0.3277
#>   -9.7967  -3.5981  11.5942  10.0331 -19.3338   7.1761  -3.4791  -0.4520
#>   -2.5461   1.9827 -13.6548  -0.1669  -4.2895  17.5683 -11.0047   4.3639
#>   -2.0913   3.0933   5.7382   1.6838   0.7033  11.7946  -5.9822  11.0849
#>   -9.4333  -7.4678   4.7480  -6.0805 -13.3319  13.9125   8.7667   2.1626
#>   -2.5727   3.0442 -13.1837  -0.4746  -8.0733   3.0046  -3.5038  10.3942
#>   -4.8250   0.9270 -10.0276  -4.3277   4.8291  20.6646   4.3883  -1.6356
#>   -4.5407  -2.0475   6.3921   7.5323  11.9548  -0.4589  -5.7074   9.3934
#>   -2.8084  -8.9770  -1.8462 -13.2308 -19.3525   8.2353  -0.3602   1.3937
#>    1.7624  -8.2600  -2.2640  -5.0981 -14.1480   8.1046   7.4925  15.0933
#>   -5.1635  -4.7577  -9.2893  -5.0399  -3.9766 -14.2620 -18.7209 -12.6348
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]