Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.3577 -11.5029   5.4368   7.4784 -12.0464 -12.9222   6.8854  -0.7405
#>    2.3401  -3.7180   3.9473   3.2565   3.0664  -0.5729  -7.8724   3.7357
#>    1.4700  -6.8309   3.7395  -8.7036  14.9630   1.5896   6.3924  -5.5527
#>    4.1134   5.5480  -7.2925  -0.6933  -9.0333  -3.2357  -2.9103  10.9224
#>    1.7250   7.5990  -7.6203   9.1139  -9.1369   9.0445 -10.6693 -16.2425
#>    0.3872  -2.4139  -1.6244  -9.5137 -11.0338  -9.4371   4.7957  -9.3803
#>   -1.7024  -5.5441   6.0816 -10.9936  -3.3187  17.9904   9.1790   6.1811
#>   -3.0576   8.6740  -7.1260  -3.9878   2.0820  -1.7184 -10.1812 -10.4482
#>    1.0684   3.5170   4.8219  -1.3901   5.8709   5.1834 -15.0297   3.9587
#>    6.4328  -6.9150  -2.4422  -0.0530   4.4362  -7.7450   7.7451  -9.8054
#>   -2.2442   4.1676  -1.3678  -7.3418  -7.8246   9.7713  11.0277  -1.5465
#>   -3.0658   6.4154  -7.2929  -2.5877  -9.3370  -1.2024   3.0135   1.5087
#>    6.2283  -3.6106   4.0987  -8.5961   4.6316   5.0416   4.5655   8.2079
#>    5.1357  -7.2991   1.7048  -7.5421  -1.3053  -5.3449  10.6260   6.8703
#>   -4.1960   7.2287  -7.8986   5.9847  -5.6804   5.2083  -7.0353   1.6286
#>   -3.4536  -1.0132   8.2458   2.7893   4.2556 -17.6390  -7.2039 -18.7607
#>    3.4786   1.5522  -1.1499  -6.0623   2.5483  -2.0332  16.5718  -0.6904
#>    3.8014  -0.7607   3.3097  12.3160   1.9188   3.3435 -19.0450   3.6366
#>    3.8861  -7.3874  -0.9465  -8.3099  14.4842 -13.2145  -7.9479  12.5233
#>    1.4569   6.9671  -8.7200   2.4446  -0.7024  14.3856  13.8013  24.0195
#>    9.1016  -6.7119  -2.3058  -1.7324  -3.6672  14.6994   3.8355  -1.4559
#>   -2.7550   6.8641   7.6980  11.6236   2.5527  -3.4397  -6.7393   1.9633
#>    1.2618  -0.3594   1.5581   6.8558  -9.2663 -24.8250  -9.1307 -12.9963
#>    4.8401   5.8627  -9.6204 -11.4485   2.2096  10.6511  -2.4471   1.1987
#>    1.7497  -3.0667   4.7167  -4.9153  -6.2847 -24.0736  -4.0647 -19.3592
#>   -9.6634  10.2246 -10.3746  11.3756   2.3961  15.6752   5.0681   1.6986
#>    0.6200   0.9469   4.1862   9.0633   9.3384   3.0812   0.2539  -4.5568
#>   -1.1740   5.2622  -5.7362  17.4633   0.5469   4.6036 -12.1881  18.7952
#>    1.5894   2.1444  -1.3400   5.1906  -3.1360   3.2803  -1.5515  -1.9465
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]