Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.5348  -8.1655 -24.7542  -8.2441 -12.5547 -12.7955   4.9513  -9.2326
#>    1.4633   4.1754   7.1138 -13.9421  -4.9384  -7.9337 -27.3915   7.8000
#>    3.1355   3.1832  -1.5954   6.3428   5.2904   5.8694  10.6204  13.4437
#>   -0.1855  -0.1931   8.7015   0.9953   6.5221  12.5236   6.2616  13.2133
#>    3.6765   3.1560  -4.2802   9.3781   6.4255   0.5003  -3.2415  12.3713
#>    4.3152  -1.9146   5.5369  -1.0773  -1.4875   4.8847   6.0725  -1.9727
#>    4.1884   2.9621  -1.4851   6.5425   4.5812  -1.7030   7.1096  -0.4215
#>    4.8322   3.3072  10.1141  -4.0470   0.7311   0.4200  -8.5551  -5.1284
#>   -2.7575  -7.7489   0.0989  -1.2531  -5.4031  -2.3388  -2.6359  -8.0937
#>   -2.4177   0.6432   0.7114  -5.5810 -11.4405 -18.0824 -13.1800  -5.2492
#>    0.5245   0.2894  -5.9346   1.5673   1.7740  -8.3037  -5.9244 -16.9341
#>    5.6190  -1.5643  -4.7206  -0.5471  -7.2404   3.3409   9.3483   2.1351
#>   -4.5992   0.0335   9.0066  -5.8400   3.0101  -1.0302  -2.3544  11.2326
#>    3.7161  -0.2010  -4.7635  10.2341  -4.6323  -6.2678   1.6322   2.8157
#>    1.6658   5.0965  -1.5648  -5.1987 -12.8074  -6.7398 -20.7075   0.6903
#>   -1.9886   0.6794  -6.6174  -5.5441 -13.7278 -19.4566 -15.5866 -31.3257
#>   -1.6197  -0.4784   5.1654  -2.0178   2.9816  21.2385   7.9521  -4.0018
#>    2.2934   2.5924  -0.5157   6.2492   4.7644  -0.8453  -5.6967  12.2277
#>    1.0418   1.5503   3.8746  11.2179  -1.4392  -9.5807  -1.1897 -12.2430
#>    1.4933  -7.7310 -11.5689 -18.9994  -8.2347   0.3438   5.1481   3.9773
#>   -0.0209   1.9486  -5.2855   0.1414  -6.8115  -5.9674  -1.0244   7.5896
#>   -0.9256  -5.1900   3.7381   4.1880  -5.8997  -0.8311 -12.4632 -17.5760
#>   -3.0319   5.3353   1.8838  10.9950  -1.5031   9.3788   4.4161  -0.6103
#>    4.4039  -2.6406 -11.7460  -0.8574   3.0945   7.8687   5.0115   8.3553
#>   -3.7699  -4.7933  -1.8307  -9.1795 -11.4717  -7.6567 -17.6715 -23.7317
#>   -2.7981  -0.2008  -6.5367  -2.8420  -9.6061   4.4547 -13.0410   0.1343
#>   -2.0279  -5.8126   3.5453  -3.0700  -7.3907  -5.6177   3.1433  -6.9872
#>   -1.1198  -1.0167 -14.7419  -2.3675  -6.6829 -19.4423  -5.3316 -10.9615
#>   -2.3683   2.2451   7.0535   4.2685   3.6669   4.6292   4.8230 -11.8283
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]