Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.1665   1.5067  -2.3744   9.7400 -19.1307  -6.6468   0.3212   4.0185
#>    3.1449  -2.3716  -1.4249  -2.7240  -7.9147   2.8236  -0.8668  -3.1533
#>    5.4766   6.0281  -6.0104  -6.4903   5.7187  -6.5033  -9.7906  -2.8765
#>    1.8407  11.9178  -2.6837  -3.8195  -3.3388   0.4482  13.9554   2.3855
#>    3.0042  -7.0564  -5.1022  -2.8988  13.4837  -2.1374   1.9346  10.9854
#>   -9.9548 -16.0084   9.6358   8.5602  17.6611  -6.6203   4.3922   6.5398
#>  -13.2892  11.9527  -8.7359 -10.9095   5.9551  -3.8451 -10.9191  -6.2697
#>    1.2874  -6.7936   0.2740  -2.6897  -9.9866   6.1321  -5.4210  17.1488
#>    9.0487   8.2448  15.3347  -2.7597  -5.1222 -19.9487  14.8469   7.5825
#>  -10.0823  -3.2585   6.5617  -0.9486 -18.7657  -3.1181  -6.1887  -1.7217
#>    7.6744  -0.1369   6.6837  -4.2367   1.1918   2.3376   4.5649  -3.8023
#>   -7.0335  -0.9418  -6.0412   6.8423 -12.2276  -2.8884 -10.8713  -5.7669
#>    1.1216   9.2916  -1.6606   6.8234 -13.0905   4.3787   5.7168   0.8070
#>    3.5010  -9.2384 -11.3722   5.8635   1.8625 -23.8648 -18.7878 -10.2337
#>   11.5434   9.5158  -7.2752  -1.8321  12.1833  10.7316   6.6226   6.6308
#>    0.1347 -11.6482   5.4031   6.9131 -11.7623   5.6156  12.0753  13.6461
#>   -3.2798  14.4603   7.0647  10.1488  -2.1011   0.7920  -7.7605  -5.3555
#>   -2.7229  -1.9100  -6.7012 -11.3806  18.4059  -6.3130  -0.2447   2.1678
#>    4.7512   1.9376   3.0951  -4.5541  -8.8287  -2.8712   5.2807   5.4567
#>   -0.2190   9.7595   0.7206  -4.5343   4.0376  -3.2648  -9.3482  -3.4055
#>   -8.3279   7.6200   9.9755  -1.6393  -4.0482   8.8089  -0.7336  -8.3496
#>    1.5301  -5.2542  -7.1282  12.0542  -6.9950  -6.8061   4.7421  13.7121
#>   -6.5954   5.1607   3.4483   7.3910   1.1823   4.2044  -2.1235  -8.7589
#>   -2.2593  -9.3590  12.2333  -2.8681  -2.1764   9.9381  -8.2455  -2.7151
#>   -7.1962   6.3508  -7.7181   8.0661  -0.5044  10.5050 -12.6510  -7.8551
#>    4.2086   1.1250  10.6485  11.2708 -11.8316   5.7033   6.1691  -2.6303
#>    9.7741  14.6838   4.4444   4.2471  -7.1187  -0.4842   6.9886 -14.4598
#>    1.8423  -3.6576   5.5099  -3.9877   6.3176  -0.4414   0.6434 -11.6792
#>   11.4376  12.5200  -1.8897  -7.8077  -8.6741  -1.8844  -7.3818   5.3271
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]