Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.9073  -0.1333   7.3165  17.6462  10.0704  -8.6331   4.3826   5.4311
#>   -5.6534  -1.3404  -0.3723  14.9100   0.1469   9.5568  20.5639  -0.5323
#>    4.2490   1.4759  -9.3294  -0.9877  10.6364   5.7173 -11.9187   2.1711
#>   -1.6273   0.9033 -17.2371  -5.0985  -5.5292  18.6141   6.3962  13.7672
#>   -2.7354  -7.6356   6.1392  14.8759  -4.5368   6.6397  -7.9194  -2.9499
#>   -4.2456  -9.4571  -3.2541  -9.6743   3.0931  -4.6012  -0.5240  -8.5434
#>   -4.8955  11.7713  -5.6417   5.4957  -0.2855  -6.8526  -4.4376  -5.2376
#>   -1.5878   0.2390  -0.6619 -11.9772 -18.7193   5.3428  -5.1760  -0.1756
#>   -1.3010  -7.7253  -0.1044  -3.0380  -2.3229  -0.6636   4.8784  -3.1704
#>    5.6073   4.4659  -8.1102 -17.6890   1.0098  -3.8680   6.2373   1.1095
#>    0.6008  -6.9828   5.0419   0.2777  -8.4427  -7.6325 -12.2256   3.2140
#>    0.6969   0.6906   8.5480  -8.7438   4.9845   8.4760   5.7264 -17.0132
#>    3.1877  -5.6924   2.1161   0.4220  -4.8705  15.4543  -0.4571  -4.5397
#>   -1.9440   2.5734   4.5019   1.1341  11.1183 -12.6034  -4.0455   0.7242
#>   -5.1935  -1.4820  -7.3870  13.0966  14.2292 -12.7091  16.9534   1.1112
#>   -1.3544  -3.3263  -5.2161   5.1820  13.6622   0.3700   1.3669  -3.2411
#>    4.7125   8.8977  -6.9927  -3.1114   4.2730   5.8029   5.4870   5.8022
#>   -1.4182  -0.4214  -0.2424  -8.6940   2.0354   6.9361   0.9637  -7.2173
#>   -2.6014  11.7884  -3.1349 -10.3876   1.9434  14.5427  -2.2284  11.4986
#>   -0.1042   2.4674  -3.5098  -4.6068   1.7206 -10.5673  -7.9448  -0.3737
#>   -2.2807  -9.1194   7.6311  -2.7139   1.6986 -13.7442  -9.4863   9.1786
#>    3.7764 -14.2443 -12.8635  19.8110  -5.1960  -3.9080   7.4986  22.1467
#>   -4.5033   2.0130   6.8305  12.8751   8.6543   3.4913   3.6773  -9.3654
#>   -0.8364  -4.2405  -1.2840   7.1787  13.5661   3.9352   6.7521  15.3517
#>    2.5675   0.3485   4.8462  -6.3982 -11.2317  -9.4855  -8.6760  -8.6510
#>    3.1749  10.9521   9.1387  -5.6469 -15.9594 -10.6227   1.5212   8.0930
#>   -4.6906   3.8005   4.3105  -5.8567  13.3828  13.2937 -10.2345   1.4324
#>   -0.5520   4.7538   7.0506   3.2671 -13.1227   9.0366  -5.7602  -3.5541
#>   -0.1539   6.2470   4.8716   1.6076 -21.1150  16.7945 -14.6679  15.4075
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]