Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.4986 -11.5616  17.7075  -5.2494  -8.6080  10.6412  -4.1486  -8.5604
#>   -1.5014   8.1685  10.1534 -10.0962  -4.1164  13.0813  -9.2264   1.6249
#>   -0.4771   6.4610  -6.7017   4.6431   5.2668   5.3868   0.5759   2.0771
#>   -3.5221   3.0362   7.6574   5.4761  -0.0266   1.6880 -10.8950   4.6914
#>    2.7918  -3.6850   2.5612   3.4621  18.7588  -7.9740   4.6522   6.1740
#>   -7.7014  -7.5810   3.9066 -15.6296  -1.6561   6.1209  -0.4515  -2.0609
#>    4.5042   4.0172 -17.0564   6.0124  -9.9312  -3.7861 -13.0459   6.2523
#>   -1.9396   2.5113   0.3493  -5.6893  -3.8686  -3.4351   0.4473  13.5658
#>    1.1565  -0.0762   4.2288   6.3359  -8.5624  24.8667  -6.3950   0.6292
#>   -0.9111  -3.3219   5.8434   5.4913 -20.2109  12.7838   6.1634 -15.3434
#>   -3.5775 -10.9472  11.9519  -2.3899  -7.6996  -8.3330   5.6682 -14.1672
#>    4.4475  -6.9260   2.1687   8.6660   5.1000   1.3275  -0.7698   5.0671
#>   -6.7633  -6.7741  -2.4735   8.2352 -11.7989   4.5448  -7.4534 -16.2284
#>   -1.9325   2.1939  12.2038  -9.9906   9.3370   9.6898   1.9806 -12.8230
#>    8.0183  -6.0970   8.5397   0.9473  12.0483 -10.6785 -14.3254   2.2917
#>   14.4963  -4.3403   4.1428 -14.3712  -0.9478  -5.3361  -0.8761  -2.9298
#>   -3.2363   7.8858  -5.9623  -2.7819  -7.1847   6.2444  10.9271   2.2144
#>   -0.8307  -0.2543  -0.9552  -2.6122 -12.6973 -10.2700  -4.1253  -2.5615
#>   -3.5312  -4.6473  -7.4007   4.4585  -1.4900   2.9080  -4.0542   6.2319
#>    3.6971  -2.3979   8.0613 -16.7626   3.0066   7.4038  11.5729   0.9328
#>   -0.1563   4.0640  -9.0012  -1.1364 -17.0368 -10.5569   0.7731  -5.2090
#>    5.3947  -0.3250  -6.7319  -1.1042  -0.2019  16.9668 -14.1842 -20.8048
#>    0.2951  -6.9340  -4.7029   0.8724   5.6602   0.1595 -14.7339   3.2817
#>    0.4051  -7.5050   4.0160   6.2964  -4.4125 -11.5116   7.1897  -3.6042
#>   -1.6231  -3.8044   4.5291  -0.3965  -7.6044   1.8322  -5.5473 -17.4043
#>   -2.0734 -14.0858   8.1246 -10.7135  -2.2392   1.0852 -13.8590  -9.7824
#>    9.0667  -6.8256  12.5174   1.1961  -6.5405  -7.6957  12.1915  -4.6673
#>    1.9290  -3.2035  10.2301  -8.1957  11.6522 -14.8859   8.5153   0.2280
#>   -1.9251   7.7923  13.2254  -0.0016  -2.0550  -3.0201   3.6709   2.1497
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]