Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.0713  -1.3891  -3.7364   2.8003  -1.2158  10.8628  -1.0642  16.6873
#>   -0.9157  -1.7584  -3.5272   1.4428   9.6808  -0.5694   0.4463  16.3610
#>   -2.7474   8.8732   3.2374 -14.9236  -2.3540 -10.2220   4.2580 -12.7886
#>   -0.7422   2.3563  -4.4488   0.7612  -2.8617   7.2342   1.5130   7.0896
#>   -1.1749   4.8561   0.2566  12.1866  -2.0537  -0.4552   1.0044  -4.9988
#>    5.8184   1.4883   1.3894  -2.8389  -6.3879   4.5360  -4.5216   1.5271
#>    0.9788  -6.6163  -3.4259   0.0899  -3.6034   8.1400  -5.3306  -2.0575
#>    0.2275   6.2984   7.8421  18.4811  -3.0549   0.6810 -14.4702  -3.9197
#>   -0.3280  -4.1556  -7.9220  -8.3568   0.3631 -21.7259   3.4446 -19.7073
#>   -2.2864   7.5275   6.3320   7.1857  -0.7939 -14.2084   7.9111   8.8500
#>    2.7593   4.6171  -5.1311   7.1520 -15.2427  -8.6611  13.1458 -10.1214
#>    2.4190  -2.0526  -1.4352  -7.5727  -1.0098  -1.4556 -15.7597  -2.1477
#>   -2.4836  -1.2107  -2.8920   1.8535   3.9960  -0.4517  -4.4619   8.4537
#>   -3.3675  -6.3678  -0.3766  -2.1243  16.7528   3.5886  -8.7334  13.4403
#>   -3.2972 -12.9585  -5.2953   1.4201  11.0337   4.6685  -6.8164  -3.5719
#>   -4.0733   4.1027   6.1862  -5.3121   9.6902 -12.3078  -6.6178  -2.3082
#>    2.6209  -0.9368  -6.3371   1.1659   7.8544  -0.2758  -2.7471   5.9636
#>    4.6108   0.9975   7.6600 -12.3524  18.6498   1.4691  -6.8994   4.2042
#>    7.2885  -1.9656  -2.5042  -2.3700   7.1467  -3.0586 -17.9357  -0.0936
#>   -1.5498   0.2952  -0.2561  -8.9250   4.5432  -5.5789   6.1423  -7.0553
#>    1.7968  -3.4078   0.5811   4.4850 -10.5317  -7.8980  12.1421 -10.6548
#>    2.9758  -3.1065 -11.4476  -0.5740 -11.3577   7.3165  -2.9953 -15.8877
#>    0.5613   1.4899  -0.4494  -3.8697  13.1069  -4.6093   8.9760 -14.6660
#>   -0.8880  -0.6265   7.6759  -2.1472   1.7095  -2.7632   7.3376  13.2335
#>   -3.9387   2.1353   0.9006  -1.3385  11.4842   2.6739   8.3091  11.4906
#>    4.8850   3.7606  11.4849   1.5844   6.6645   3.1988  -1.1656  -0.4946
#>    3.9589   5.0266   5.8077   1.3242 -14.3829  -6.2456  10.3693  10.4500
#>   -2.3239  -6.3055  -4.6844  -0.7084  -4.3852   7.1450   4.4544  11.8290
#>   -1.6436   8.1255  -6.9389   6.5739  -3.2645  -2.3038  -4.9236  12.4564
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]