Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.2252  -0.3885   3.1973  -3.5339   2.8147 -16.3260   4.0746  10.0773
#>   -4.8157  -5.4095  -7.6057  -2.4680   1.4776  -1.1189   3.6877  13.0026
#>    2.0277  -8.4288   5.3565   2.1133 -15.0941  -9.2588   9.1088   1.2556
#>   -6.0163  -9.0147   0.9696  -2.1023  -6.7009  -0.2769  -9.8304  -1.0884
#>   -1.3823   8.2539   6.8940   5.0474  -6.7468 -16.2439   5.5660  -3.5269
#>    4.1907 -12.8203   6.2572   6.0169   5.5718   5.6370  -7.9526  -7.1754
#>    7.7503  -6.8204  -2.1092 -11.7831   7.1270  -5.5607  -7.3347 -22.2215
#>    2.6685   5.3079   7.3130  -0.5005  -2.4824  -1.1010  -2.7851   0.6030
#>    7.2903   1.8689  -1.7594   2.8831   4.8618  17.9360   7.0525   7.6044
#>    0.2720   4.5557  -4.5180   5.6481 -13.8854   5.7550   0.5561   6.5487
#>   -4.7828   7.1625  -0.3585   6.8535 -16.3417  10.9967  -7.2570   3.8019
#>    4.1288   3.9307  13.1714   4.0660  -7.6507  -1.8858  -7.0124  -0.6186
#>    1.9427   2.0075 -15.2395  -4.6477  13.7427 -14.4464  -5.9937  -3.4752
#>   -0.8762   8.3348 -10.8578  -6.2108 -10.7122   0.5910 -14.7635  16.5812
#>   -1.4663  -0.6813   5.3343  -4.1554  10.3532   7.5616  -2.9899  16.8423
#>   -2.9359   5.0038 -10.9314 -14.4882  13.2078  10.6742 -12.4362 -11.5545
#>    9.8864   3.3392   9.2816  -8.2919  -2.0573 -12.9057  -6.2794   0.3458
#>   -3.8619  -2.4551  -0.6039  -2.9226   0.7176  -0.8125 -10.6692 -16.4457
#>    8.0827  -3.0954  10.8343   1.4278  -2.7254  -4.0521   3.9515  -8.1328
#>   -4.8420   1.2645   0.6693  -4.5070  -4.1695   3.5016  -4.8062  -6.3979
#>   -9.8054  11.8688 -11.0467  -0.6891  -1.4587  12.5690  -5.3828   4.2436
#>    4.0169  -4.1838   9.4294  -6.3963 -22.2871  -3.4767  10.3169  -7.9913
#>   -3.8602   0.7468  -0.0239  13.7182  19.0105   5.0947  -2.9130   5.2912
#>   -4.5866  -3.8352 -13.9920  -1.8625   8.4516  -0.5557  -9.4444   7.3761
#>    3.7303   3.3662  10.1199  -4.2078  -6.4704  13.6834   2.2981  -6.1724
#>    0.9836   0.6583   6.7852  -5.0240  -2.5403   8.0557 -11.4654   3.2339
#>   -1.7957  -6.0365  13.7216   9.0109   3.2689   4.1396  10.0480   1.8485
#>    0.4922  -0.6920  -5.2689   5.5823  -1.5781   5.2839   5.2519   2.8843
#>   -2.7528  -6.7819   8.8660  -4.8329  -1.8585  -5.0525 -11.5975 -13.5229
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]