Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.1421   5.1005   6.3756   3.0242  -1.7890  -0.0752  -9.4592   9.5929
#>    9.1886  -2.5094   3.7383   1.5203  -0.1215  -2.4680   6.8985   7.6141
#>   -3.3726   0.0491  -1.6166   1.5534   2.1055  -2.5977   4.5761   0.4785
#>   -3.0613   2.5231   4.6744  -7.9230  -1.2539   4.2176 -19.0129  -0.3671
#>   -6.4955  -0.4935  -5.4609 -11.3029  21.8231   5.4475  -0.1270  -3.4324
#>    3.8240   8.1776  -9.7088   4.7063   7.8201   4.0441   7.7109  11.8174
#>   -1.3794   3.7247   8.6044  -1.1092   0.2287   2.8196  -1.0421  -4.3983
#>   -0.3584  -3.0243  -6.1558   3.1619  -2.9773  -4.8625  -6.3427   0.0480
#>   -1.8318  -7.4239  -5.9552   3.2584  -1.1009   9.6435  -9.5529  -0.5400
#>   -2.0592   0.9040   6.2600   4.5080  -0.2265  14.8222 -19.5682  -0.4773
#>    3.3633   3.0818   6.8357  -4.9976   1.4043 -12.6591  -1.1324 -14.9365
#>   -1.3426   3.0267  -2.5215  -5.6055   4.0473   6.1949  -0.8557  -2.7664
#>    2.9531  -2.6276  -1.7268 -13.7289   2.6638 -11.3466   1.4789   0.1456
#>   -2.6737  -1.7386  -0.9397   6.9001   3.2668   0.3635   4.6864  22.7694
#>    2.2905  -2.3735   1.3030  -6.7019   0.2906  -2.9821   9.8220  -6.6338
#>   -1.6710  -3.3310  -7.9693  -2.0831  15.9339 -21.2978   0.0316   5.2226
#>    1.9985   0.5449   6.9291  -0.1135   0.8711  -8.8468  -9.0955  -1.2982
#>   -2.5539  -3.5041  -7.8758  11.2184  -4.3667   9.7960  -5.4079 -10.3765
#>   -3.1110   2.7042  -0.7801   4.8717   7.2892   6.1529   0.2941  -2.2046
#>   -1.4013  -0.3707  -4.4843  -1.8269   6.3329   1.0388   8.1716 -20.9029
#>   -1.5555   0.8025  -5.2178   8.6754  -8.8652   1.8154  10.0706   2.0052
#>   -1.0192  -1.6109   1.3900   2.5270  -1.4131 -16.9873   6.9295   4.0839
#>   -3.0345   0.9880   5.6847   1.7235   8.1257  -0.3313 -10.8991  -7.1830
#>    1.2044  -3.3022  -9.3236  -9.6966  -6.0113  -0.5471   6.7196  -0.4600
#>   -0.9289  11.2173  -9.4773 -10.1015 -10.5779   6.9870  -9.9863   2.5563
#>   -3.7346   1.0427  -5.0693  -0.3007   7.6441   1.2593  -8.2914   1.9771
#>   -3.6842   2.7304   5.0856  -9.0035  -7.8967  -5.4815  -1.7939  -9.0919
#>    8.7726   0.5162   6.9417  -1.8542   6.0750 -17.7315 -15.4733  -0.0509
#>    1.5246  -5.6712  -3.2559 -14.5112   0.0824   8.6666  -4.1288  -9.8026
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]