Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.1184  -3.3914   3.5386   4.8204  12.9131  -3.9902  -3.6834   2.6504
#>    2.0937   9.9643   0.6152  -8.1275  -3.8642  -6.2577  12.8017  -3.7366
#>   -4.7885   6.3748  -6.9688 -10.1528  -3.1594  15.6162  -0.4162   2.9919
#>   -3.7028  -2.2968  -1.8115  -0.1669  11.3491   0.6841   5.9650  -7.2097
#>    1.2183  -7.4212  -9.5517  -2.2135   2.5497  11.7977   4.4066   1.6804
#>   -2.4640   5.5718   9.0832  -5.2969  -6.4107   4.7070  -3.4104   5.8539
#>    1.2236   4.6811   4.3700 -10.2066  -1.8688  13.0760  -8.1619 -12.9306
#>    3.2344  -5.1901  -5.5404   9.3484  -1.3643  -5.8200  -5.6616   3.3928
#>   -2.8054   8.8666   0.2990   0.8359  18.6408  -8.6254  10.8603  -7.5724
#>    3.2662  -0.9177  -4.2459  -4.5522  -6.7637  -3.7180   2.2466  -4.9650
#>    2.6177  -5.1616   1.7978   0.8093   4.6547  -7.8625   0.5825  20.5200
#>    0.0521   4.6744   4.3752   1.1111  -9.5294 -16.2558   6.3788   8.0982
#>   -4.0264   4.0318  -4.1553  -2.7115  -1.0242   3.4481  18.3130  -5.8286
#>    0.8374  -3.6940  -1.9500   4.1537   7.2160  -7.2355  -0.5481   6.0047
#>   -1.8729   9.5238 -11.4821   5.0693 -24.7403   0.8790 -11.4123   8.0887
#>    6.9491  -6.0570  -9.9921   0.9248   8.0352   8.8198   1.9148 -11.2405
#>    3.6265   5.3439   1.9402   5.8628  -0.5520  -5.8001  -1.7733  10.2100
#>    1.0978  -2.1936  -3.2103  -6.8701  11.6005   2.9078 -15.0036   6.2479
#>    0.4065   6.3179  -5.0924  -1.9448   4.2506 -11.5814 -14.3717  15.5185
#>    1.4303   0.7157   4.2281  -4.6103  -1.8024  -8.4511 -17.1149   7.3381
#>   -2.2034  -6.0149  -1.5405   3.0604  -1.2579  -4.9353  13.5781 -15.8534
#>   -3.1694   1.2324   4.9319  -7.8790  -3.1080   2.9640  -7.5862 -15.6985
#>    1.4229  -1.3193  -1.3312   3.4420   3.9482  -9.7694   6.7131  -0.0102
#>    1.0884  -6.8790   1.4889 -11.6752  12.4900  12.8057   0.5308   1.5974
#>   -3.6419   3.2762  -1.0193  -8.0901  -5.5657  -1.2578  -1.6740   3.5809
#>    4.8395   1.8088  -4.3352   2.9404  -8.3270   9.2320  -6.9363  23.3426
#>    6.3950   3.9892   0.9364  -7.4454  -2.5684  -9.5110  -0.5148 -12.5684
#>    4.7665   1.8504 -11.4966  -1.0155   0.0305   8.9580  -5.8364   9.1137
#>   -1.7245  -2.9653   2.5804  -8.5640  -3.6396   2.7732   8.8802  -6.4961
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]