Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.3882  -5.6375   3.3337  -4.2113  -5.1220   2.1206   2.9429  -6.4123
#>   -2.0929   3.7979   4.8818 -10.7826  -0.7399   5.6565  10.8957   0.0442
#>    1.1564   9.3124   2.1425  -7.1434  16.9385   9.5808   3.4169  -7.6975
#>    2.0162  -4.8358  -7.9028   9.4568  -7.6777  12.9684   0.0543   0.7349
#>   -6.5962  -5.6523   5.7711  -9.5944  -5.2324   1.0229  -5.6332  -4.1196
#>    1.6091   1.6447   4.3775  -7.9018  -1.5888 -10.9845  -0.4826   7.2167
#>    5.6023 -11.5480   9.8741   2.8806  -2.4897  13.1150   1.5476  -8.9286
#>   -3.8666  -4.1480   6.9449   3.7074  -3.6567   0.7162   7.8136   6.7056
#>    3.9344   7.4954  -2.2658   5.0120  14.5782   9.6941 -24.1259  -3.8134
#>    3.4736   4.6094  -6.4786  11.0541   4.9624  12.0381 -16.0136   9.0893
#>   -1.2453  -0.2281  -8.1431   9.0818 -15.3424  -3.1134  -5.9119  -6.7203
#>   -2.8442   2.8682 -12.2221  -0.0253  10.7853  -8.3045  -2.9308  -1.7246
#>   -5.3292   1.0662 -13.6011   8.5612  -4.2575 -16.8734   9.5213 -14.4156
#>   -4.5860   8.1660  -8.7316  -2.5354   3.9093   7.4462  10.1490  -8.9896
#>   -1.2647   0.2898  -4.9297   0.1718  -4.9453 -10.9422   5.7885  -4.2872
#>   -0.7879   6.8248  -9.1792  -4.5692 -12.1961   3.1089  -1.4162  -6.5928
#>   -4.3046   4.6135 -12.1633   2.1282  -6.7270   7.5190 -10.5536   4.5035
#>   -0.9244   5.3456 -10.0481  11.6947  -4.2357   3.8381  -0.9333 -13.1780
#>    0.2396 -10.4122   8.3905  -2.8289  -1.7785  -2.4726   5.6739  -1.8931
#>   -2.7638  -3.2207   0.5264   4.1840   9.3212   2.1217   4.8761  -9.1915
#>   -5.4288  10.4635  -6.6288  -4.1944  -0.2254  12.2724  -8.7880  -6.4964
#>   -1.6099   2.3348  13.7276   2.8682   6.1364 -11.2073  12.0459   6.0707
#>   -1.0823  -2.4767  17.1690   7.5757 -12.1324   8.4062  -5.0866   2.2977
#>   -2.0806   5.2303   6.7068  -3.4549   2.9727  -3.4938   7.9442  -0.0951
#>    2.8569   1.2620   3.3952   5.2807   5.1862  -5.9284  10.4194  -6.0755
#>   -0.5321   3.9184   4.9964  -5.9221  -2.0684  -2.1772  -0.4218   4.0627
#>    3.2842  -0.7853   1.0195   4.3084   9.6259  -7.9292   3.2390  -8.3738
#>   -3.3307   8.0498  -6.9351   4.9217  -5.0553   5.5116   6.5875  -5.1922
#>  -11.6466   4.1149  -6.1882  -5.0388   2.8018   5.9636   5.4360  -4.6984
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]