Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.5769   3.6999  -6.0659   1.8950  18.8089   2.9882   0.1701  -9.6516
#>    2.3300   4.8127  20.8212  -2.9309  -7.5499  12.7341  -2.7778   1.2416
#>   -0.9553   0.3224   4.8024   7.8902  17.1435 -18.6646  -6.4648  -3.8409
#>    1.7988   0.2368  -2.3640  -2.3082  -1.5501  -3.7152  14.1127   4.4126
#>   -5.1518  -3.7266   9.9275  -2.9645 -19.7677 -12.6179   3.9635 -23.1221
#>   -1.6550  -0.2157  -9.2024   1.3442   9.4048  -5.5418   2.3322  -2.1767
#>   -1.6078   4.4288  13.2789   4.2840  10.6254   6.9888  16.6138   4.3396
#>    2.5804   6.1802  -3.1520  -8.1541  23.6013   5.5769 -14.4109 -13.1500
#>    7.0583  -1.2802  -1.5059   0.6221   5.1654   0.4849   0.0845   6.6796
#>   -1.4325   3.4901   4.9616  10.4807 -11.6857   2.2027  -5.5280  -5.8913
#>   -1.6947   5.6898  10.1728   0.6328  -4.1815  -6.6796   1.1801   0.8540
#>   -0.3089  -0.7125  -4.7085   4.4590  -9.6947   8.1522   8.7146  14.6710
#>   -3.0787  -0.7186 -15.6195 -10.7061   7.9491  -9.7064  -7.2294 -16.2145
#>    5.5673  -3.9517  -1.8552  -7.8654  12.9373   0.8053  -7.8160   1.8945
#>    0.6084  -0.6048  -4.7930  -5.3370   0.0875 -14.5758   9.0340 -12.4234
#>    4.0741  -2.2310  -2.4606   5.2816   6.8725  13.4219  17.4930   5.4563
#>   -5.6137   4.0506  -3.0975  -6.8762  13.0790  13.5825   2.5790  -0.9329
#>    4.1005  -7.7916   0.9554  11.2041  -0.9825  -1.8771  16.1312   6.3327
#>   -6.3347   2.9383   1.1525  11.8107   0.0532 -18.1470  -5.3442  -8.3375
#>   -4.2295   2.2828  -4.3911  -7.6017  -7.9426  -4.8084   5.2738   0.0644
#>    0.8727  -7.5921  -7.3066   2.5335  18.9390   8.0480  -2.4516 -12.5345
#>    3.2530  -5.2204   3.9687  -5.0461 -10.9118 -11.1129  16.2837  -8.9751
#>    9.0317  -0.0355  16.4196   8.3034   7.4787  17.7903  -2.6614  -9.8352
#>    1.7456  -0.1923   1.7843   5.7171  -9.4774   4.0547   4.1175  -2.0968
#>    3.4511  -6.5052  -0.7408  -5.2430  -1.1230  -0.3486  -6.9488  -2.5491
#>   -4.7732  -2.2171   9.3527 -18.5265  -9.4010   4.5676 -23.5528 -14.5930
#>    2.5849 -14.7275  -3.4327   8.8043  -1.3967  -9.3132  19.5402   2.7114
#>    0.8586  -5.4971 -11.6360 -12.0045   7.2131  -2.5833   2.8064   4.1965
#>    2.5169  -3.6094  12.4799  -9.8692  14.2700   1.5547  -5.9614  -6.7219
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]