Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -3.9251e+00  1.0507e+01  2.4230e+00 -9.3884e+00  1.0578e+01  1.7638e+01
#>   2.9297e+00  3.9893e+00 -5.4870e+00  4.1940e+00 -1.6327e+01 -2.4909e+00
#>   2.3776e+00 -1.1970e+01  3.7766e+00 -2.7342e+00 -7.0404e+00 -8.5575e+00
#>  -3.1634e+00 -9.2432e+00  1.6460e-01  4.0039e+00  8.7190e+00 -1.9020e+01
#>   1.6439e+00  1.2032e-01 -1.1187e+01 -2.7550e+00  1.4795e+01  1.2625e+01
#>   4.1004e+00  1.3276e+01  5.6931e+00  3.4047e+00  1.6519e+01  3.7577e-01
#>   5.6441e+00 -9.3118e-01  3.9413e+00 -8.4335e+00  5.3580e+00 -9.3067e+00
#>   2.1816e+00 -7.6274e+00  2.7862e+00  6.3485e+00 -2.1919e+01 -3.6023e+01
#>  -1.4643e+00 -1.7915e+00  1.3256e+00 -4.2151e-01  2.6725e+01 -1.1901e+01
#>   3.3523e+00 -4.1394e+00 -6.6715e+00  2.7542e+00 -1.3253e+01 -5.6085e+00
#>   1.3707e+00 -6.1247e-01  2.2904e+00  1.4232e+00  9.4121e+00 -2.8460e+00
#>   3.8085e+00 -5.7799e-01 -8.8938e+00 -5.1401e+00 -1.0353e+01  4.9632e+00
#>   5.9334e+00  5.6984e-01  2.4416e+00  9.0712e+00 -9.8821e+00 -1.5963e+01
#>  -4.2479e+00  1.7568e+00  3.7153e+00 -1.7799e+01 -4.5830e+00  3.9147e+00
#>   2.4343e-01  1.1014e+00  3.1471e+00 -6.5222e+00  9.7521e+00  3.4463e+00
#>   2.4217e+00 -1.4588e+01  1.6516e+01  3.7279e+00 -4.6441e+00  2.9028e+00
#>   7.3945e-01 -4.8969e-01  1.8647e+00 -5.0440e-01 -3.9538e+00 -4.9888e+00
#>  -1.1126e-01  1.7495e+00  8.6850e+00 -7.5299e-01 -7.3366e-02  5.0347e+00
#>   3.3965e+00  1.4069e+00  2.8055e+00  1.7341e+00  1.8828e+01 -5.5847e+00
#>   2.0342e+00 -6.7599e+00 -9.9024e+00 -1.4267e+00  1.4068e+01 -7.6219e+00
#>   3.7510e+00  7.6361e+00  8.6025e+00  4.6924e+00  1.2183e+01  1.4577e+01
#>   1.2813e+00 -3.0837e-01 -2.4280e+00  9.9777e+00  1.3731e+01 -5.5644e+00
#>  -1.3464e+00  3.6263e+00  4.4396e+00  3.1731e+00  7.2427e-01 -5.1732e+00
#>  -5.7910e-01 -5.3597e+00  1.7859e+00  6.5198e+00 -1.7551e+00 -2.0532e+01
#>   1.0587e+00 -8.1541e+00 -7.9228e+00 -1.2270e+00  9.3116e+00  5.6032e+00
#>  -4.2042e-01  3.5258e+00 -8.4417e+00  7.8183e+00  4.1683e+00 -3.9511e+00
#>  -3.5359e+00  2.4812e+00 -6.4805e+00 -1.7652e+00  2.6884e+00 -1.4403e+01
#>   3.9462e-02  3.2904e+00 -1.8477e+00  5.6478e+00 -1.7134e+01 -1.6245e+00
#>   3.9892e+00  1.3241e-01  3.9840e+00  5.4257e+00 -8.1077e+00  3.8294e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]