Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.5053  -5.2882  13.1655   7.6257   0.7011   1.8434   2.0876  -9.3783
#>    1.6084  -0.3284   4.4167 -10.2272   1.7189  -6.3495   0.9291   3.8253
#>   -1.2417  -5.2576   6.6428 -15.9543  -0.7636   3.9436   1.5620   7.0781
#>   -6.6054  -4.6395  -2.7569   5.1535  -1.9909  -0.4658   8.5227  20.8417
#>    4.3429  -2.7913   0.6601   5.5923  -1.3976   3.0620  -9.9278  -7.2373
#>    1.1962   1.0849   8.3127  -1.2397   4.9367 -15.2281 -12.5882  -1.8689
#>    4.2994  -5.2082   6.6265  -1.5751   2.7990   2.0422   5.1783  -9.8674
#>   -3.6049  12.3046   3.3047  -1.4357   5.5301  -7.1007  -5.3464 -12.7144
#>    0.0490  -8.6226  -6.0132  -1.8902  -6.3303  -3.8688   5.1013  -7.4052
#>   -5.2068   0.5364  -0.3414  -5.7948  -2.1157   6.4512  -0.7927   9.7658
#>   -2.3647   1.5635 -10.2313  -2.2308   6.9851  -4.3007 -14.9460  -6.6010
#>    7.1191  -7.7357  -2.8662   1.2441   6.8581  -5.0777 -12.1840  -8.8488
#>   -3.1558  11.7749   2.5635  -3.7280  -2.1844  -8.4337   4.3109   2.6282
#>    5.9018   2.1669   5.7530   6.3210   3.5055 -11.4038  -1.0771   0.9966
#>    0.6842  -0.7764   2.2240  -5.3535 -19.5333   5.8139  16.1249   7.0070
#>   -5.5895  -3.9846  -2.9154  -2.3600 -11.9460  -5.1393   9.7134   0.5279
#>   -5.2018   7.6652  -8.1905 -15.1997   3.5457   0.9614 -13.4560  13.0019
#>   -0.0909   0.7115  -3.6647   1.3892  -0.9921  -8.5601  -6.3127   2.7011
#>   -2.1656  -3.4334  14.0818  -4.2657 -16.5150  -0.5265  -9.7821 -15.5044
#>   -1.3348   3.3338   2.1280 -17.2628   0.1597   3.9819   2.6876  15.6737
#>   -2.9587  -7.0131   4.0506  13.5727 -11.2502  -0.1138   3.9749  -4.3252
#>   -2.8363   7.9332   0.1713  -0.9838   6.0600  10.8869  12.9432 -17.1466
#>    2.6943  -1.1877  -6.2822   2.3910  -1.7580   5.3719  -4.2724   9.4830
#>    0.0289   4.0011   0.1762   4.5121   8.7467  -1.2181   4.3418  -9.8587
#>   -7.6979   4.6446   8.6255   7.3321 -13.4193   1.7853   5.9917 -22.4304
#>   -2.3164  -7.3921  -2.2722  -1.7125  20.1809  -3.1852  -5.9433 -11.2801
#>   -4.2642  -1.5430  -2.4964   7.3038   0.2109   3.2611  -5.4978  -9.0477
#>    3.9575  -7.6769  10.8411   7.3051   9.5547 -15.1213   9.6025   5.9013
#>    3.0948   2.1284   2.8115  -5.2922  -8.1585  -9.3831  -7.9480  -5.8921
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]