Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  3.3158e+00 -1.3733e+00  3.1595e+00  4.3987e+00 -4.7746e+00 -9.5638e+00
#>  -3.3810e+00 -3.0883e+00 -9.6490e+00  1.4027e+01 -2.1209e-01  7.7574e-01
#>  -1.9686e+00 -1.0042e+01  2.8313e+00 -3.5157e+00 -1.7434e+00 -4.4888e-01
#>  -2.5470e+00 -4.7182e+00  1.2976e+01 -7.0974e+00 -1.0395e+00  1.0415e+00
#>   6.6187e+00  1.1019e+00 -5.8131e+00  1.0304e+00 -5.5213e+00  1.0914e+01
#>   2.3268e-01 -9.1717e-01  9.7365e+00 -2.8710e+00  8.5599e+00 -1.0572e+00
#>  -3.5751e+00 -6.6213e+00  1.5835e-01  1.0955e+01 -5.5852e+00 -2.7782e+00
#>  -6.7882e-01 -9.1570e+00 -2.6060e+00  3.5995e+00 -8.0370e+00  3.9905e+00
#>   3.6805e+00 -7.8985e+00 -1.2258e+01  6.9025e+00 -1.5569e+00  8.2934e+00
#>  -2.0527e+00 -1.0640e+00  1.2439e+00  1.3316e+01  1.0130e+01 -5.1673e+00
#>   5.0703e-02 -4.4768e-01 -3.7236e+00 -7.8727e+00 -1.4327e+00  9.1184e+00
#>   2.2481e+00  5.8095e+00 -5.4465e+00  9.5230e+00 -7.0585e+00  7.6294e+00
#>  -3.8083e+00 -1.5424e+00  7.7481e+00  4.0491e+00 -1.1616e+01 -3.5068e+00
#>   3.0340e+00  5.6851e+00  2.7603e+00 -6.1869e+00 -2.6774e+00  8.6502e+00
#>   5.4263e+00 -7.2491e+00  1.5371e-02 -4.0232e+00  3.2012e+00  3.8026e+00
#>  -2.2558e+00 -3.3827e+00 -2.5688e+00  9.5761e+00 -7.9289e+00 -7.7552e+00
#>   3.5321e+00  1.0572e-01 -1.6686e+00  1.6287e+01  1.0543e+01 -1.8844e+00
#>  -1.6559e+00 -4.7794e+00 -9.4186e-01  7.4016e+00 -1.3248e+01  1.4378e+01
#>   7.0699e+00 -1.0054e+00  5.2837e+00 -7.6112e+00  1.6362e+01 -5.1468e-01
#>   3.5995e+00 -7.6611e+00 -6.1763e+00  2.1641e+00  7.5329e+00  5.4029e+00
#>   4.4697e+00 -7.0573e+00 -2.1081e-01 -1.6110e+01 -2.2277e+00 -5.0545e+00
#>   3.1488e+00 -3.1915e+00 -9.3440e+00  3.4594e+00 -7.6778e+00  1.2790e+00
#>   4.7184e-01 -5.9112e+00  8.2781e+00 -7.1073e-01  8.8135e+00 -4.9935e+00
#>  -3.1300e+00  4.6730e+00  2.7499e+00  5.3732e-01  6.6814e+00  8.5965e+00
#>  -6.5879e+00  9.7054e+00 -7.3337e+00 -4.9858e+00  6.4609e+00 -6.9166e+00
#>   4.1149e+00 -1.3010e+01  5.2348e+00 -1.2110e+00 -1.0244e-01  6.8535e+00
#>   1.1397e+00  4.7392e+00  4.1739e+00  4.6337e+00 -8.1101e-01 -2.0126e+00
#>  -2.6700e+00  4.7758e+00  5.8475e+00 -9.4536e+00 -1.1948e+00  2.4427e+00
#>   6.3983e+00  2.0782e-01  2.8006e+00  1.8527e+00  6.6680e+00  3.0064e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]