Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.5302   7.4609  -3.1585 -13.7527   5.1273   8.2301   4.6096   6.6931
#>   -4.9888  -1.1592   5.9946  -6.3133  15.2317 -11.1088   0.8558   0.9007
#>    1.1157   0.6483  -5.0992   1.6953   7.1417  -1.9133  -7.1007  -1.8449
#>   -1.8132   5.7408  -0.4385  -8.4599  -7.3236  -5.4762   4.3241 -11.0101
#>   -1.1766  -0.0095   4.4126   0.9369  13.8020  -1.7208  -5.7929   2.3152
#>   -0.6789   7.9608  -6.9852  -8.1544  -4.1623   9.9244   2.0482  -0.7753
#>    5.8390  -3.9679   9.4883  -3.7941   5.9773   1.6224  -5.0175 -10.2250
#>    6.5897  -2.0061  -0.4780  -2.0321  -1.2115 -11.0127  -6.9186 -11.5038
#>    4.4335  -1.9312  -3.6561  -3.6964  -9.2487  -1.8504  -1.9235  -2.4285
#>    1.1906   0.4218 -13.1158  -2.7462  -0.6947 -14.8137 -19.0172  -6.0532
#>    1.0849  -5.6861   8.3693  -5.0957 -12.1714  -0.2178   3.6752  -9.4546
#>    3.2231   3.4549   0.8942   0.1351 -12.8427  -3.6838   6.9495 -15.9270
#>   -4.9737   5.0249  -3.3337  -4.7019   0.0038  -5.6428  -4.7861 -15.5297
#>    0.3398  -3.8754  -3.6809  -0.8704   3.0986   7.3265  -2.7212   2.1693
#>    1.5634  -4.7796  -4.2754  12.4478  -5.4022  13.2408 -10.6108  -5.5946
#>   -3.0200  -4.4312   7.3137  20.0564 -13.3734  -4.2761   0.8917  15.2086
#>   -1.7346  -8.0061   4.3595   3.0647  -6.7908  -2.0382 -14.9234  13.3441
#>    4.7716  -6.8492  -4.1255  -3.2965  -5.0936  12.8586  -3.7150  -9.6911
#>    3.1345  -6.2557  -1.0610  -6.6504  -8.3376  -7.1763  -2.8598   7.6896
#>    1.8594  -6.2265  -5.4548   8.1998   1.5762  -0.2868   6.1741  -3.1890
#>    7.8698  -4.6599  -6.3080   5.7471  -7.7260  -9.2000   0.9692   1.3625
#>   -1.5551  -3.9833   3.2558   8.0024  -1.0470   7.5045   6.1477  -6.2405
#>   -5.6728  -3.5148  -0.6022  -6.2364   3.9148 -10.2815  -1.9691   8.8284
#>    1.9917  -8.3895   0.9485   5.5995  16.3556  -0.3060   2.5721   4.5412
#>    1.2072  -1.7830  -3.3803  16.5771   2.7804  -0.9298 -14.4621  -5.9201
#>   -1.8567  -0.1805   7.8073  -6.3938  -6.2755   2.6891  -2.7914   7.3244
#>   -2.4147   0.1031   2.8027 -11.2535   4.8145  -0.6065 -13.4533  20.2361
#>   -5.3423   4.6456   7.8144  -0.2465  10.5469   2.1239   6.2948  -3.3863
#>   -2.9631  11.4709  -7.1773  -2.2108  -4.6594  -1.8901   6.9905  -2.9681
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]