Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.1291  -4.8646  -4.2891  -9.1528  -4.7428   8.8296  -9.4110   5.9907
#>   -3.2813  -1.2503   5.1839   4.6711  -1.3453  -8.7234 -13.1278  -8.6858
#>   -1.2731   2.4527  -1.6127  -6.9756  -1.7658   3.7570   8.1094  -2.4726
#>   -0.3758  -0.9501 -11.9411   6.7601  11.3844  -3.5608  -5.2607 -18.0350
#>    2.1768   3.0810  -8.9570   0.1833  -6.7498  -2.2198  -4.7497 -11.5652
#>   -0.6538   0.1368   0.4215   1.2811  -6.3397  -3.4008  -9.7296   5.0212
#>    1.5566  -3.4377 -11.1473 -23.1401   6.6897   9.7476  -8.9070   0.2547
#>    2.1590  -5.3496   1.9577  -3.1005   0.4391  13.2163 -12.9520  10.2757
#>    3.4163   0.5404  -5.4523  -1.4999   6.3129   4.5915   5.0885 -22.4102
#>   -0.1500   4.7165  -6.3969  -0.5013  -6.8271  -2.6151  10.3491  -1.9686
#>    0.1385  -4.5128   7.1454   8.4980  -7.6024  -3.2456  -8.2200  -5.5244
#>   -0.8311  -1.5435  -3.5417   8.2156   4.5247 -17.4467  11.4453  -2.2970
#>    0.0354  -0.6774   2.9414   3.1997 -12.6237  -0.0639   8.7506 -16.2478
#>   -0.0917  -4.4242 -11.5557  10.9201 -18.2688  13.8708  -1.2152  -6.0835
#>    2.9319   0.9511   0.2457   1.7022  13.3937 -31.9332   5.0905 -14.5450
#>   -2.5685  -0.6687  -6.4727   2.4827   3.4920  -4.2061  -1.7277   6.2278
#>    1.0346  -0.0187  -0.2808  -7.0874   7.5189   4.3680   9.1987   0.1042
#>    0.2346  -2.3141   4.2194   8.4285   9.0052   4.1065  -0.7846   4.6152
#>    0.6630   6.3299   4.9150 -16.5484   9.6567  -5.3148  -0.9744   1.9677
#>    2.2043  -4.7269  -3.4019  -4.7103  -7.4737  -5.8915   0.8374  11.9170
#>    1.7877  -0.3949  -5.5115   1.7419  10.4402  -0.8515  -3.4744   0.1583
#>   -2.1111   0.9639  -2.7522  -1.5198   8.5625  -5.1654  -5.2597  -7.5880
#>    6.3709   3.8323  -3.7914   6.1575   6.9214 -10.7268   0.7086 -14.2755
#>    3.1491   0.4865   4.0189  -5.0522   3.8979  12.1324   9.7154   1.6511
#>   -2.5973   4.4947  -2.7172   0.8901  -7.3440   3.1856  -5.8841  -0.7590
#>    1.6833  -6.1188   7.3933 -11.0124   8.6053  -1.1891  -7.5639   9.5085
#>   -2.7147   5.6757  -1.4409  -1.5626  26.2187   4.0817 -10.0181  -2.1445
#>    1.0811  -1.8778   0.0458  -1.1885   1.2260  -6.1626   2.8077   6.3606
#>    0.1529   2.7195 -11.4865   8.7827  -1.3166   8.7460   3.5551   6.6428
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]