Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.9634  -2.1189   6.9933   6.7551  -1.2731  13.6416   0.9391  -8.4009
#>    0.9585  -4.8098   6.3080   6.6735   1.3885   8.4027   6.9740   8.7249
#>   -8.5176  -5.0136  11.5027  11.6370   0.9878  -0.0652   6.3883   5.9150
#>   -2.0541   1.9331  -5.2989  -4.2279   3.0294 -17.4898  -8.1258  -8.0669
#>   -1.7593  -3.0666  -2.9338  -4.1924  11.3178  -1.9572   3.9425   7.2994
#>   -0.1125   2.9847   1.1755 -13.9048 -13.1871  -0.0271 -16.2216  -4.7269
#>    0.9938   5.4095   2.1474  -3.2914   0.4252  -7.6682   2.0662  -4.0125
#>   -6.4356  -2.6723   9.3741   3.9924  -2.9319   1.0455  12.7100   3.7649
#>   -2.7448   7.4272   7.1932   0.5552  12.6281 -11.0734   2.4735   2.9999
#>   -2.4677  -1.1992 -11.1228  10.1555  -5.6388  -5.2955  18.5261 -21.5554
#>   -4.1135  11.6634  -0.1527   1.5652  10.8241 -18.3431  -4.1136  -2.1168
#>   -0.6431   3.7420 -15.5412  -6.9537  -3.1427   1.6307  -2.8625   6.9647
#>    1.1862  -6.5815  -2.6717   0.7721 -13.3258  24.1550  -3.0700  -2.9531
#>   -5.2664   5.1404   7.4783  -1.1585   1.6324  -8.6374 -11.7270 -12.2003
#>    2.4854   8.3930   7.6577  -1.8265   4.7067   1.9423  -0.7939   6.8646
#>    1.0867   7.9611   9.3926  -2.6108  -2.7309 -10.1290 -12.5139  18.6703
#>   -3.2532  -2.2800   1.6769  15.4416  -3.2000  14.3978 -15.9769   7.7059
#>   -8.0695  -0.6125  12.9881  -7.0819  10.2731  13.4905 -13.0871   6.5810
#>   -2.2851  -1.0061  -4.0493 -12.5716 -17.4049 -16.6185   9.7586 -19.6113
#>    6.5563  -6.1149   6.6585 -19.8784   5.4551  -2.6953  -1.5559  17.5781
#>   -0.3738  -1.7223  -3.4185   5.6411  -3.5979   2.3300  -6.5011  -3.5088
#>   -5.2876   1.6845   1.8581  16.0508  -7.9941  -2.4060  10.6341   0.9863
#>    2.2257 -10.0559   4.2528 -10.6350   8.4570  -1.6204  10.6364  16.6823
#>   -0.4783  -0.3859  -5.6143  -3.7442   5.1978 -19.0043   4.6475 -13.7237
#>    2.3551   4.0041  11.3602  13.2078   5.6834   3.3409 -10.0502   9.3571
#>    1.0158   2.8700  -7.0669   5.4621  -1.4123   4.2548  19.0174  11.4263
#>    1.0228   1.6800  -0.3683  -6.7928 -15.0823   0.1189 -15.7798  -0.1358
#>   -7.7267  -3.7833 -13.6235  -2.4546 -15.5665  -1.3207   0.1238  -0.1484
#>    0.1599   4.9029   0.7132   1.4927 -10.8345  -4.1904   0.7673   3.0012
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]