Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -5.2854e+00  3.6748e+00  1.1340e+01 -8.0060e+00 -1.8533e+01  1.8022e+01
#>   6.6399e-02  5.5108e+00  1.1812e+01 -1.4636e+01  9.0455e+00  6.2777e+00
#>   2.0567e+00  3.2526e+00  1.6891e+00  9.7628e+00  3.4888e+00 -1.4080e+00
#>  -4.4788e+00  6.7970e+00  8.4392e+00  5.8214e+00  5.0826e+00  7.4525e+00
#>   3.1812e+00 -1.3239e+00 -4.4516e+00  9.3512e+00 -8.1573e-02 -9.7498e+00
#>  -1.8557e+00 -2.4615e+00 -3.5828e+00  2.9450e+00  5.5229e+00 -2.8928e+00
#>  -1.1589e+00 -1.0524e+01  2.0079e+00 -1.1473e+01 -6.9318e+00 -2.0957e+00
#>   3.3323e+00 -2.1964e+00 -2.2083e+00 -8.6160e-01  4.4608e+00  4.5596e+00
#>  -2.8078e+00  8.1108e+00 -1.2004e+00 -9.8161e+00  3.1329e+00  4.5608e+00
#>  -2.2928e+00 -2.2071e+00 -7.9694e+00  1.5612e+01 -1.1740e+01  7.8730e+00
#>   5.7446e+00  5.8287e+00 -2.6056e+00  6.5024e+00  1.7215e+00 -3.9471e+00
#>   6.4998e+00  5.4711e+00 -2.0538e+00 -8.0152e+00  2.0071e+00  5.0637e+00
#>  -2.7832e-01 -1.1710e+00  6.8870e+00 -1.8351e+00  7.3008e+00 -8.4586e+00
#>   1.2370e+00  8.9925e+00  1.6610e+00  4.7886e-02  8.6614e+00 -6.9745e+00
#>  -4.5987e+00 -8.1394e+00 -1.1540e+00 -2.5100e+00 -1.5579e+01 -2.7475e+00
#>  -3.0322e+00  1.7446e+00 -3.5399e+00  1.7055e+01 -1.5016e+01  1.9871e+00
#>  -1.2929e+00 -7.2360e+00  1.6254e+00  9.0959e-01 -9.1212e+00  3.2200e+00
#>  -2.2908e+00 -2.3980e+00  7.2231e+00 -4.9012e+00 -4.3715e+00 -7.2722e+00
#>  -3.9062e+00  7.4249e+00 -4.5046e+00  8.8493e+00  8.1219e+00  9.2568e+00
#>  -3.9148e+00  1.7247e+00 -3.7426e+00  7.5013e+00 -1.4422e+01 -3.2874e+00
#>   1.2136e+01  2.1706e+00  3.0333e-01  1.0357e+01 -1.0914e+01  4.3608e+00
#>  -2.5630e-01 -2.8257e+00 -3.7106e+00  1.7088e+01 -2.1276e+00 -1.0641e+00
#>   1.7301e+00 -4.8668e+00  1.2625e+01  9.5681e+00  7.6758e-01 -4.6250e+00
#>  -2.0617e+00 -3.6791e+00  1.0802e+00 -1.0535e+00 -8.1750e-01 -1.6113e+01
#>   4.3720e-02 -6.1010e+00  6.2978e+00  3.9806e+00 -1.0816e-01  1.1081e+00
#>   2.3214e-01  3.2767e-01 -5.4881e+00 -1.1842e+01  9.8497e-01  2.0633e+00
#>   1.4826e+00 -1.3188e+00 -7.4861e+00 -1.7610e+01 -1.4766e+01  6.6494e+00
#>   2.9644e-02  6.1235e+00 -8.1032e-01 -3.4601e+00  4.8536e-01  5.5695e+00
#>  -1.6765e+00  2.6809e+00  1.2888e+01 -1.0581e+01  9.5134e+00  8.0104e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]