Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.3084  -2.6527  -6.6690  -6.2388   7.9745   0.1387  16.1290   1.3354
#>   -0.4008   1.7436   1.1758 -16.9155  10.1861  -3.5534  14.6896   7.9829
#>   -3.9088  -8.6229  -4.0162   7.0910  -0.0656   5.6821 -17.7087  -9.9054
#>    0.2856   2.9937   0.7262  -7.3576  11.7551 -13.0053 -13.3563  -4.8922
#>   -0.6249  -3.1469 -10.7416  -2.7572   3.1370  -7.0292   5.6554 -16.7614
#>    0.8849   3.5734   3.6515  -3.8351  -2.2469   4.8754  -9.4825   5.5913
#>    0.8742   0.7249  -4.8684  -1.2590  -3.0129   7.5406  -3.6718  -2.5581
#>   -1.2690   2.5827  -0.1430  -4.3769   2.8786  -0.9139  -9.6525   4.2509
#>    5.4319   0.1980   2.5245   1.5419   5.0326   5.2488  -8.2841   3.1551
#>    1.8893  -7.2337  -4.2040 -11.4066  12.8453  -0.2538  -8.7042  11.3699
#>   -1.4561  -1.9339  -3.0159  -6.5123   4.6182   4.5038   1.3489   8.0682
#>    2.4118  -0.6938 -11.4867 -11.9825   4.9728   3.0988  -1.0993 -15.6856
#>   -1.5690  -1.7284  -5.5715 -10.3462  13.8564 -10.3143  -0.6245  -3.6578
#>   -3.1689  -4.1262   6.1911   1.9108  -0.4782  -7.2009 -10.2976   6.4599
#>    2.1155   6.0198 -11.3150  12.8724   0.2410   3.6052  -4.4510   1.4278
#>   -2.7984  -2.4024  -4.1913  -6.5829   5.0755  17.1415   0.8979   8.1381
#>  -10.3394   3.5031  -5.3673   6.7157  -9.6838   6.0128  -5.2478  -7.5787
#>   -2.2845   2.1419  -2.6732   3.8418  -5.7591   0.1244  -3.1656   2.3932
#>    1.4321   6.6675  10.9740   5.4758   9.2386   6.3516  -8.3053  -0.8762
#>   -7.3632   0.2381   1.1655  -4.0988   6.3909 -12.3532   5.8355  -3.8087
#>   -3.8399   1.5311   0.7651  -2.5867   0.6334   9.2457   9.3058  -1.2372
#>   -4.0431   3.4887 -11.0139   3.9407   1.0414  -4.7170  -0.0943  -8.4672
#>    4.5022   8.7033   2.5863   4.1716   1.7553  -5.6598 -12.8303   5.3894
#>    1.2814  -6.0862   5.6521   2.5104   2.7557  -1.7143   0.1416   9.9364
#>    2.8988   2.3338  11.9772 -18.7349  -0.3752   9.7266 -11.7764  -1.2736
#>    3.4933   0.7896   5.5317  -2.7346 -14.8673 -18.3191  10.5994  -8.1838
#>   -1.3410   1.6668  -8.9429  -2.4570   0.8294  -0.5653   8.6265  -6.9138
#>    5.9540  -2.4910   8.1585 -10.5760  17.7183 -10.7642  21.9741   7.6669
#>    0.3996   0.3715   2.3548  -5.0259  -9.5933   0.6241   5.3651  -5.3829
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]