Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.6401  -2.5831   0.4159  -7.1520  -7.2351  -2.1578   0.4835  -8.5173
#>   -0.3331  -3.1325  -6.5741   3.9996  -4.4162  -0.9210  12.0622 -15.2074
#>    3.6572   4.7206   1.4199 -13.3515   3.5247   1.6424  -2.2507   6.4456
#>   -4.4097  -6.0304  -1.6463   6.3406   7.1989   7.3262 -19.6877   7.6660
#>    0.9425  16.7110  13.7050   8.7852  -2.1834  -1.1728  -0.3365  16.4781
#>   -2.3153  -6.7202  -2.8509  -4.7249   1.5607   4.9574  -5.6735   2.1833
#>   -2.5378   5.6889   5.0566   1.3673 -14.9998  -4.7035  -4.7804  -1.0703
#>    1.9003  -1.8971   3.2476  -2.4027  -4.5505   7.5628  -3.3978  11.7650
#>    0.8261   3.9782   1.8204   2.7564   4.6250  -7.1479   2.4218  -3.0522
#>    0.5334  -9.1953   2.0166   2.4491  -8.6064 -17.4266 -12.4698   7.0287
#>    5.8036   2.5122  -3.9728 -15.1829  -0.5533  -0.9827  17.9853  -2.5163
#>   -5.9285 -10.1173  -7.8766   4.7093  -2.5732  -2.4022  -5.8615 -17.3463
#>    2.0807  -0.8532  -8.1681  -5.1181  -2.4538   8.9166   5.6505  13.0549
#>    5.2863   2.8604  -1.8468 -10.2875   7.2477   1.9058   3.0549  -7.5895
#>   -1.2321 -19.4230  -8.0292   3.7972  18.5011   0.3350   1.2894 -11.0238
#>   -3.8372  -1.9769 -10.5129  10.3450  15.7427   6.8228   7.6640   2.2408
#>    0.8242  -9.3311  -8.5199  13.9306   1.0770  -8.6603  -3.0465  -3.8877
#>    0.3273  -3.5149  -3.6275   0.1986  -4.7611   2.8715   5.6782  -0.1657
#>   -0.4391   5.0174  -1.1923   9.3298  -7.6348  -9.3681   0.9813  -5.3831
#>    4.1964   2.0078  -0.2064  -5.3143  -9.9290   3.8044  18.0066   7.1000
#>   -2.6017   0.8339  -3.5465  -1.7419   5.7017  13.3537  -0.5728 -10.1605
#>   -1.7625   4.0556   6.2917   3.5699 -10.1929   3.3468   2.1582  10.1212
#>    0.0480   2.8004   2.9130   7.1557   7.2072   1.1319  -3.7761 -10.0867
#>  -11.2661 -10.5068  -4.2567  -1.4616 -13.4523   3.5889  -8.1411   0.4031
#>    6.8797   2.8287  -2.8715 -11.8776  -5.8335  15.2102  -0.0720   5.1954
#>   -0.6723   9.1666  -2.9684  14.1851  -4.5928  -2.2186 -13.7532   0.2759
#>   -1.5048  -0.0565   0.5430  -7.9637  -1.7286   8.8616  -0.2807  -7.8359
#>    1.7720   4.8721  -2.8726   8.2193  -5.8137   7.8720   1.7807   0.4034
#>   -5.9229  -0.8515   3.9625   7.2085  18.7395  -9.7820 -14.9163  -0.2697
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]