Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.1486   0.4180   2.3362  10.1355 -11.6547  14.5680  -2.7744   7.0221
#>    0.3726  -6.7273   5.1815  -2.9318   6.9910  -2.6249  17.5577  -3.1474
#>    4.6627   2.0331   6.3895   5.0477   8.0689   0.8428  10.0353 -13.2123
#>    1.7250   6.2686   4.7684   3.3660  -0.3008   6.3923  18.8653  -2.9119
#>   -0.9508   5.4184  -8.3759  12.8622   3.2476  -2.1658   3.9886  -6.6234
#>   -5.4366   2.8506  -2.7577   2.4579  12.0660   2.9783   0.1621  -1.0438
#>    0.8846   2.8051  -5.3151   9.2090  -1.4776   4.8012   8.9492   4.6521
#>    2.6094  -0.1162  -1.8671   8.2082  -1.7676   4.4484   7.7561   8.7118
#>    0.2587   1.3496   5.5574  -9.4317   7.6389   4.5458   3.6058  -7.1111
#>   -2.2140  -0.5732  -0.6215   1.4629  19.9422  -6.9431   5.9533   2.8313
#>   -0.0512  -9.6142  16.3305 -13.1875  -5.7998   2.3187  12.4877  16.8112
#>    0.7588  -0.7800   0.9903  -3.0100  -1.0004   1.5260 -13.0391  12.5577
#>    1.1992  -0.9008   0.5791 -15.0567   3.3570   1.0584  11.9344  -4.3042
#>   -1.0914   5.2032  10.6245  -3.6338  17.8898   1.1691  14.5160   0.8610
#>    3.0452  -8.3594  13.8541 -15.2025  12.6859   9.2150   7.0997   4.6608
#>   -2.0162   1.9312   2.3540  -9.1483  -4.7802   4.2845   4.6646   6.1065
#>    0.8148   2.6860   6.0630 -13.0293  -4.9549  -6.8598  17.8879   3.0443
#>    1.8003   6.0941  -0.1190 -14.9523  18.4398 -18.0376   8.2901 -10.9122
#>   -0.3461   0.8070  -6.1487  -0.4203 -18.3168   2.7885  -1.9943  -7.2176
#>   -3.1263  -2.2884  -4.1138   0.8780  -1.9174 -15.4835   2.8614  -9.6799
#>   -1.6875 -11.0200  -1.2485  -8.7929   7.8574  10.3119   4.5025   0.0466
#>    1.1639  -4.6049   4.4523   3.1320  -7.5961  18.0624  -1.1348   4.9271
#>   -1.2993   1.9691   6.2381   9.8365 -16.6143  -1.9288  -2.1846   4.9015
#>   -3.6649   6.3857 -12.6690  11.4589  -7.7623   6.3265   0.7082  -9.2483
#>    1.3833  -1.5041  -2.5956   0.8632  -4.6065 -10.2501  -6.9011  15.3251
#>    3.0918   2.5469   4.8063  -1.6920   1.8477  -6.1251 -14.2310  13.1755
#>    2.0301  -3.7078  10.6239  -3.3186  -0.0901   1.8053  -1.8888  -7.9343
#>   -1.7024  -5.5715  12.1888   1.5251 -10.9475   3.5056   8.1607   6.9929
#>   -4.3960   0.2280   5.0461  -6.8970   7.4780   2.5366   6.0431   1.5565
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]