Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.5241  -6.6105  -2.5098  -1.1631 -11.9459   2.8617  -8.1429   0.8602
#>    1.4371  -1.9466  -3.3223   0.3914  -5.8611  -1.3755   4.0936   4.5257
#>    1.1848  -8.2921   1.5414   1.8680   2.1891  -9.1567   4.2429  -6.4962
#>    1.2868  -2.6563   7.0895  -1.5914  -3.0932  -5.0522  -5.3872  12.4971
#>   -0.0636  -0.1494  -5.8263   7.7692  -2.2397   6.2112   5.2084  15.8812
#>    1.5759  -2.9412   6.4825  -3.6117   0.2703 -13.0258  -7.9352   8.0975
#>   -8.3086   0.4081   1.0629  -4.6033  -1.2748  -4.4130 -18.0365  -6.1526
#>   -3.1405   3.5757  -6.7192   3.4812  -3.6206  -3.7001   0.9971  -9.9844
#>    0.8902   6.5601  -7.3067  -7.2048 -10.5320   0.9201  -0.3789  10.8338
#>   -1.5847   3.1060  -1.9632  -5.2784   1.1243 -18.6504   0.3773  13.1150
#>    0.8029   2.0276   1.7008   2.5495 -12.5190   2.3717   3.0033  -7.4787
#>   -2.5035   2.3720  12.1657  -5.4723  -2.0615  13.0120   2.6952   1.7027
#>    3.5616  -5.2663   4.9291  -3.9708  -9.8986  -5.3730   1.3208 -11.2106
#>   -0.5526   3.3468   7.5704  -5.0245  15.5968   5.0745  -4.6227  -4.1672
#>    0.2581   2.6052   2.2878   4.9090   5.8177 -14.9042  11.0765  -3.6337
#>   -0.5347   7.9836   3.7807  -9.9688  -5.8295   6.1420   8.0543  -8.2888
#>   -0.5956  -6.9765   1.6121  -2.4186  -1.9474  -3.2551  -6.3230   1.8781
#>   -3.8463  -0.2261  -5.1651  10.5434  17.7098  11.4023  17.3140  -5.6790
#>   -0.8691   1.7434   7.3259  -4.6593  14.4725 -12.0757  13.2988  10.8886
#>    3.4773   0.9981  -3.3264  -5.7048   1.6356   2.9326  -3.8091  -0.7470
#>   -4.1989  -1.0227  -0.0806  -0.1918 -13.6849  10.8462   8.4267  10.2687
#>    2.9518  -5.5131   0.3660   3.6415  -0.6280 -12.8127   9.7201  -5.0915
#>    2.4034   1.3897  -4.9980  -2.9005  -8.9922  -1.8948   6.7102  -6.1329
#>   -5.3484   4.8152  -6.2878  -8.1571   5.7738   8.2518  -8.7443  -9.1058
#>   -1.8031   1.5618  23.2834   7.2027   1.2619  29.4484  -2.2196  -8.4747
#>   -2.0249  -9.0140   1.5120 -10.4450  16.5195   0.9498  -2.2485   6.6265
#>    1.5159  -0.8630  -2.2518   2.7770  -1.9740 -10.3201   5.4000  -0.6376
#>    2.3152  -1.6225   5.2797 -13.8286  -1.3343  -1.0486  -8.2868  -9.1731
#>    6.2860  -7.4401   5.4883  -4.4536  -4.1213   6.9910   3.1885  -0.2608
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]