Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  13.7947   2.5338   0.7632   2.9036 -16.4623   2.8739  -6.6868  15.1166
#>   -1.0689  -5.9770 -16.8500  -6.4321  -2.5601  17.3383   1.4085   3.3700
#>    8.4413   0.5606  -5.2854  14.2811   3.7404   8.7945 -16.9414   6.7419
#>   -1.1295  -7.2339  -0.8085   1.8396  -0.2952   4.6844   3.6479  10.3714
#>    0.8030   4.1979 -13.5038  -0.3537   8.2806  -5.9175  16.0079  -8.0565
#>   -2.5288 -13.5963   1.6865   7.6406   0.1958  -0.5540  10.6520  -9.3769
#>    2.2874  -3.3749   8.3840  22.4714  -0.9150  -4.5015   2.0626 -15.3220
#>    9.1034   2.3429   3.0618  22.6502   1.4818  -7.1789   0.1254  -7.6601
#>    0.5493   0.4801  -5.0182   6.6577  -9.3151  13.2020  -0.9649  13.2152
#>    5.4726   6.5334 -11.3417   5.3896 -14.7750  -7.5358  17.3621  15.9466
#>    1.0110  -0.1055 -14.5246 -11.1808  11.2463   8.4237  17.5939  -2.9944
#>   -0.2152   6.3014  -4.7559   3.7997  -5.1239   8.6246 -12.0295   6.7328
#>   -7.4599   6.1428   1.4298   6.5173   2.2786   1.1433 -13.1787   1.7906
#>   -7.1422  -3.0805   3.6707  -7.5718   3.3439  -9.0189  12.9944   8.3639
#>    4.4710   5.9336   9.2558  -5.7281   7.4405  -4.4779  -9.4223   4.9088
#>   -6.6440  -1.2451   2.1036   0.5100  11.2198  -5.4621   7.7630  -4.8327
#>    5.9850   6.6062   3.5664  -0.1662  -0.0934 -19.5731   8.6600   2.2670
#>   -2.0238  10.0851   8.9234 -11.0557  -3.1175  -4.0966 -14.3831   3.3075
#>    1.7487  -2.0622   1.1507  -3.6502   2.5151  -1.4237   0.8828  -5.4385
#>    3.0589 -15.2477   2.1287  16.4558   3.2919   3.0283  12.2381   1.5679
#>    4.3980  -3.1104  -1.2222   4.9559   2.6995  -0.8031  15.1102 -13.5079
#>    0.0195  -7.3096   0.9886  -3.2651  -9.8428   1.3861 -11.9131  -8.5184
#>    0.8830   5.6842 -14.9174  -5.2692  -6.7949  -8.1011   4.6173   3.6564
#>   -5.0815   4.1582  -3.9113  -1.9125  -6.5908  -3.3058  13.2561  11.8925
#>    0.1451   9.8560   5.8319  10.3825  -8.6928  10.5888 -19.1213   0.1267
#>   -4.5363   4.1310  -9.3549  13.0102  -6.1510  13.5969   3.9595   7.8782
#>   -4.0576   6.4734   1.8888 -13.4919  11.1995  -5.2824 -10.3169  -1.0373
#>   -4.2521 -15.9634  -6.5217  17.2514  -2.2173   4.2139   3.2634 -11.1617
#>    2.7222  -1.6951 -16.7538   9.3292  -0.1484  -2.2809  -3.2573 -11.9074
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]