Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.1263  -3.2995  -6.8499   3.4250  -4.4949   2.0901  19.7979  11.6216
#>   -3.4225   6.8055  -1.8742   0.3810  -2.7873   1.7575  10.1109 -17.5283
#>   -1.0326  -2.0154  -2.6185  11.2922   7.0120   6.0341   4.7335 -17.3300
#>    0.0090   5.4714   5.2369   9.7761   6.8921  -0.6261  -8.4179  -4.2226
#>   -6.9387  15.1107  -7.6862   8.6296  -3.2747   5.1086   4.8436  -1.3108
#>  -10.7100   0.2438  -7.9833  -6.3129  22.3129   1.5404  -8.1553  -5.7745
#>   -1.6942  -3.7610  -5.2611 -10.6192   6.4236 -16.3267 -15.0147   7.5569
#>   -1.8896   0.1632  -8.8816   1.9832  -4.6610   3.1822   2.4175  -3.5194
#>   -0.8180  -1.6764   7.9434   8.8746   0.1882   7.9637   0.5682   0.1319
#>   -3.6120   3.6529 -21.7427 -17.2403  -1.3196  -0.4262 -10.8854  -2.3179
#>    1.0045  -8.5426  -1.5386   8.8221   0.1740   4.8425   7.3495   3.5047
#>   -8.0602   1.5164  -0.3133  -8.1582   4.8397  11.4651  -0.9355   6.7210
#>    2.5050  -1.8752  -3.8295   3.9642 -14.3427   2.3876   5.8475   9.1228
#>    4.8871  -8.8593   6.0681   9.8904 -12.7820   8.7316   8.2740  -4.5386
#>    2.5105   3.1735   2.9833   1.1737  11.6105  -1.4447 -10.1210 -12.4892
#>   -3.8079  -5.2227   5.7367 -10.5731  -0.0578   9.4709   6.7843   6.8760
#>    0.6187  -3.3159  -6.5572 -12.8097  -4.0680   2.8317   2.7580 -14.2941
#>    1.6378   3.5106  -7.0904   7.7901   4.1549  -7.2848  16.5020  -7.0322
#>    2.2497   1.9743  -2.0776  -1.4534 -11.0633  -6.2913  14.4168  -0.9427
#>   -1.4921   1.8053   2.7699   3.1624   2.6228   3.8708 -11.3816 -19.2984
#>    4.9077  -1.3091  -6.5809   2.9601  -5.3835  -4.3104  -4.7364  -3.7259
#>   -4.8035   6.3365 -10.6861  14.4819   3.3127   6.5760  11.0356 -14.0739
#>   -1.9715   6.3672   9.8981   4.0823 -17.2995  -1.6230 -19.7748   6.7292
#>   -3.5003  -0.6278  -4.8581  -7.4844  -2.2637   5.9898   4.9102   5.2662
#>    2.5669   0.0606   0.4553  -2.5592 -14.2362  -1.3208  10.0025  12.0042
#>    3.3075  -4.7069   2.2256   3.8358  -5.2105   2.0164  10.1147 -13.6650
#>    6.2596  -5.5477  -3.7203   5.6011   1.6133   0.6583   7.1832  -8.6746
#>    0.2670  -4.7093  -7.3341 -14.9006  15.8819  -0.0140   8.5994  21.3920
#>    7.4221   2.5603  -0.8745   8.7119   7.6064 -12.7218  -9.4746  -2.3883
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]