Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.4763   3.1991  13.4774  -3.4858   2.2561  -5.1967  -1.0032 -13.6544
#>   -0.9017  -1.4695  -3.7822  -5.8128  12.2709 -15.5512   4.7333  -2.1362
#>   -1.5191   1.7340   9.0895   7.2467  -7.6598  -0.5333  -5.8253  10.2957
#>    0.6257   1.6699  -7.1565  -1.8950  16.9727   8.0054   1.7757  -0.9106
#>    4.3899   4.8305  -9.5993   4.7532  -5.2090  24.2837  -4.6637   1.5770
#>    6.1429  14.0785   0.6504  -9.4999 -13.6346   9.6830  10.4116 -12.7820
#>   -2.4023  -6.2946 -10.5812  14.8514   3.2357 -22.2641   7.9937 -12.6888
#>   -5.8200  -8.2939  -1.5581   8.6414  -2.1435  12.9897   3.1354   4.2942
#>    1.8354 -13.1413  -7.2452  -2.6975   2.8653 -20.5293   1.0366  -8.3956
#>   -1.3844  -2.4053 -16.1340   5.6926  11.2205  -5.4909  -4.3766   3.4237
#>   -3.3057  -2.4136   3.9815   0.5405  -8.0092   5.8690 -11.7245  10.0896
#>   -0.3799   6.5983  -1.6820  -2.2447  -7.8894  12.7640   5.1222  10.6850
#>    3.8552  -0.6239   9.5985  12.5973  14.4742   0.3462   5.0355   6.6388
#>   -5.3456  -3.5007  -3.1671  -8.9408  -9.8173 -11.6317 -17.7961 -19.2570
#>    2.6505   2.0982  -1.6179  -7.6340  -1.1005  -8.0356  16.7230   0.8080
#>    6.5362   1.8656  10.6968   0.1784  -6.4580  -6.3944   3.2904  -0.0460
#>   -2.0874  -8.8064 -10.9271 -10.0385  10.5621 -15.1165  11.4228  -3.4608
#>   -4.7603   5.2273 -16.9864  -9.3303   9.0765  -1.9792  -8.8475 -10.0268
#>    3.1716   4.4086  -4.5863 -14.7082  -7.7026  17.0936  -6.3686  -2.3549
#>    3.0688  -2.3278 -14.4320  -5.9807   5.4149  -1.5947  15.6839   1.4522
#>   -1.1321  -6.3695   1.4049   4.1897  -5.1423   7.2334  -6.6373  -6.0481
#>    1.1893  -1.6950  -2.7960   1.1850   4.2834   9.0558   8.2781   9.3085
#>    1.2608   4.4506  10.3867  -7.7154  -7.2089   8.5517   8.4216 -10.9405
#>    3.5575   0.0223  -0.7728   5.7968  -2.3291   7.3460   6.0459  10.6768
#>   -1.9043   2.7975  -5.3748   9.8314  -9.2043  17.9851  -4.0373  -8.1495
#>   -0.1140  11.4338  -7.4941  -5.1621   1.0061   1.9352  15.7178   7.4022
#>   -3.0814  -7.1198  -0.1821  23.4073  -2.6656   9.9200 -12.0106  -8.9327
#>    3.2421  -0.4603   2.7208  -3.2017  -2.7707 -20.7351  -3.1152 -15.4678
#>   -8.2086  -0.1619 -14.1866  -6.4864   7.1313  -1.5980  12.1451   0.5984
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]