Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.8932  -2.0713  -4.0737   4.1315   2.2760  -1.5096  -5.7657   9.6791
#>    3.4355   5.5413   7.9894  -6.5664  10.3926   3.1793   9.4783  -5.7267
#>   -0.3436  -1.4552  -8.7192  -1.0733  -2.7217  -3.1607  -0.5358   0.9727
#>    0.0727   1.9369   1.7837  -9.1050   4.5387  -6.3431  -6.3885  -5.8546
#>    0.2551  -3.9216  -1.6940   0.4776   8.7969 -13.2725  -4.4309   3.4779
#>    2.4638   0.7834  -0.7341   6.6022   1.0181  -3.0592  -7.6316   9.1312
#>    0.9941   3.3839  -3.0296  -5.4311  -7.5928  -3.8566   2.4822   1.4926
#>   -3.8666  -5.3484  11.3141 -13.7623   3.4270  16.5873  -5.0307   9.7007
#>   -5.5150   1.0962 -10.4088  -2.5994 -14.2867  -9.7989   4.3456   2.2776
#>    2.3738   3.0231  -3.9812   5.5009  -2.9664  -6.5303  -7.0118   0.8565
#>   -5.6846  11.3316   7.5426 -10.0873   5.4816   9.3499   5.6659   0.3846
#>    1.0948  -0.2139   3.0683   5.2087  -6.6109   8.9807  -5.0749   6.9469
#>   -0.7407   4.2562   2.5403  -7.2324   3.7875  -6.4142  -5.3062  -7.7208
#>    0.9383   8.6473   2.0267  -0.2182   4.5350  -0.1330  24.8931  10.6752
#>    3.8673   2.9442   1.7187   7.8546   5.7454   2.0289  -2.0806  -1.6416
#>   -1.8768   3.6114   1.8394  -0.6235  -8.5282 -13.5414   4.0460   0.5257
#>   -2.0652   3.1778   9.4194 -13.8667   3.8779  -6.7669  13.2777 -10.7672
#>   -2.6346   1.5405  -5.5058 -12.4514  15.2564   0.6358  -7.6318  -3.3670
#>    2.1646  -0.7044 -11.7646  -7.5702   2.6579  -4.4330   5.5109  -2.1172
#>   -6.2654   7.5425  -3.2070   4.2264  -6.7046  -3.1195  -0.1177   1.4326
#>   -1.7840  -1.2662  -1.1029  -1.4421  -6.7131  -0.7028  13.4651  -7.1891
#>   -0.4560   3.8835  -9.0556  13.5493 -17.0315   3.1390  -3.8692   2.3350
#>   -3.3990  11.9149  -2.0294   3.4183  -3.4599   1.3274   3.4518  15.7664
#>    0.5956  -2.3161  -6.4232  -5.8835  -7.4787  -5.0783  -2.0488  -0.3241
#>    2.1024   1.2587   5.1429   4.1633  -5.7656  21.5015  -3.1320   5.6243
#>   -3.6716  -1.6576   3.5248 -10.1290  -5.3994  -2.2684   5.3945  -6.6079
#>   -6.7262  -1.7769  -3.6115   4.1797   5.0108 -14.9885  11.5037  -4.8517
#>    1.2102   2.8119  -2.3591   1.5180  -3.5208  10.5050  -1.4352   8.3344
#>    1.9722  -7.3622  -7.7818  -0.8147  -7.9097   4.2222   9.0620  -2.3745
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]