Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -6.6381 -11.2781  -1.3455  -7.9717  -2.3995  -5.1908  -4.0357   3.0083
#>   -2.0488   5.3050   4.3383   7.2691   4.7083   7.2177  10.5858   3.1618
#>    3.2360   1.7198 -12.9560 -13.0169  -7.8034   4.4029  12.5074   1.5965
#>    2.3492  -2.8854  12.1951  -6.5052   5.8412   9.4313   7.7753   4.4446
#>  -12.7125  12.4923 -13.2631 -23.9653  -5.6913 -25.6804   2.4361  14.3362
#>   -0.6402  11.9804   6.6849  16.3526   2.2180   4.0535   5.3524   6.4587
#>    0.4203   7.7779  12.5793  -6.8537   5.7280  -1.2263 -12.6303  -4.0681
#>   -8.6661   7.4700  -7.3675   1.8260  -4.5078  21.4076  -5.7726 -10.8265
#>   -1.4525 -10.3781   5.2280 -13.9898 -25.8349   0.3306   2.5590  10.1028
#>    6.0421   4.0709   6.1298   4.9057  14.6863  -1.5021   3.1425  -3.5812
#>   -4.1433   2.3130  -3.1723 -12.7480   6.4497  -7.1629  -2.5255  10.2038
#>    1.9153   3.5078 -21.0605 -28.6907   7.0422   7.9271  13.3959  -0.1937
#>    4.4448  -0.2499   2.4584   2.8119 -12.3500  -4.5219   0.1855  -0.9240
#>    0.6268  -1.8480 -15.5198  -1.4687 -10.9551   6.6434  -1.0625 -16.7800
#>    3.2624  -2.4835   3.2675   1.7680   3.6428  10.1497   8.8498   4.7802
#>    2.1873  -8.8770   9.0112   0.7410  -1.8937  -3.4301   4.1375   9.1288
#>    8.8393  18.1114   1.3926 -14.9816   2.1591  13.1254   7.0856  -9.7541
#>   -1.9617   1.2498   4.7559   9.7897  -2.3320  -5.8991  11.5919  -8.5876
#>   -3.9569 -14.3071   4.6844 -22.6041 -10.7498   7.6222   3.9912   9.0001
#>    0.1471   2.4011 -14.7976  11.0447   2.1268   6.8767   7.2419   3.5667
#>   -3.3723  -1.2168  -2.9682   7.6631  -3.7299 -13.6346 -13.1171  -1.3177
#>   -0.5858  -8.3778   3.1610   2.4439 -12.8671   3.7458   1.6639   8.2608
#>   -1.6083   4.5758  -9.3536  -4.8745 -19.6471 -12.1164   5.7199   4.0614
#>    7.4932  -3.2957 -10.7993   2.5713   0.3202   8.4239  -9.6792  -8.0543
#>   -3.7637   3.2352   6.8229  13.8301   8.9963 -21.4470  -3.3607  16.5821
#>   -0.2795  -0.5072   3.6197  24.8879  12.2167 -15.0070   5.0200   9.6308
#>   -9.4968   4.2249 -15.1089   3.3890 -10.1411  -5.3799   0.6304   6.6949
#>   -6.3144  -0.6407   0.5219  -6.4050  -6.5410  -2.9859  28.8673   1.8785
#>   -2.0376   2.6181  -3.1998  -7.8830  -1.9127  -9.5849  -0.9238  -1.0547
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]