Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -6.5961   1.3479  -5.4424 -10.1965  -9.1203  19.7759   4.0871  10.3732
#>    4.0015  -3.9523   8.9241  12.6199  -6.2270  -0.2374 -21.6610  -1.2145
#>    2.6113  -0.3334   2.7446   8.0010  -1.7605   6.9716  15.8759 -16.3283
#>   -1.4677   2.4171   6.4093  -2.9188  -5.9023  -3.1372  -1.0510  -4.6193
#>   -1.4512   5.1175  -6.0271   2.6268   0.1876  -4.5978   0.7993   1.0577
#>   -0.3704  -1.2681  -1.9332  12.5689   2.6007  -4.2864  -2.5611   7.8388
#>    1.3626  -4.4484   0.3373 -12.0716  -0.2901 -14.6026  23.2427   9.3474
#>    2.9290   5.9176   9.3835  -3.4849  13.0877 -11.9134  11.6285  13.6985
#>   -2.3670  -1.3382  -2.6629  -2.0832  -3.2460   4.0115   9.2915  -4.9740
#>    5.2405   0.5141   2.9430  -6.1215   6.3309  -6.3233   5.8485  -3.0343
#>   -3.9727   1.7415  -3.2438  -1.2571  16.1495   1.9045   2.4323  20.8059
#>    4.6956   1.4049   6.0273   8.2532   3.3817   4.1919   0.7491  16.5394
#>   -0.7243  -1.6894  11.5470   2.9442  -0.3963  -9.1095 -22.3322  10.9116
#>   -2.3077   7.6460   5.8251   3.6346   5.7371  -7.4354  14.0567   5.2460
#>    9.1488  -1.4202  10.5841 -10.5222  18.3329  -5.3599   1.5699   0.4653
#>    0.1174   5.5929  -3.7956 -12.6185 -16.8088  -5.5735  -8.5341   3.5093
#>   -3.1056  -2.9416  -0.3845   3.3999   6.4145  -6.2847  12.9265  12.7150
#>    5.5353  -0.6794   5.0698   0.4425  -3.6062 -16.7604  -1.5712 -15.2956
#>   -1.4479  -4.4894  -2.7139   1.7283   6.3405  16.2618  -1.3049  -0.5835
#>   -1.6050  -8.0857  -9.7073 -15.4501  13.4671  -5.4352   0.1391  -4.0753
#>    0.2362   1.8896  -3.3139  -0.9341  -3.0765 -13.2700 -12.0395  20.9739
#>    4.2724  -5.0254  -3.4486  10.8656  -4.3449   0.2690   2.5012 -14.5849
#>   -8.8865   3.1888  -1.0096   0.3684 -14.9980  -5.6859  -6.2278   9.3406
#>   12.0837   0.4254   8.0898  -1.5672   5.3778   2.5958  -3.4677 -15.5494
#>   -0.1140   8.1386  -8.3447   2.6901 -12.2824  -8.6623  -7.7759  -1.7154
#>    4.4374  -1.1778  10.0092  -1.6295   4.0312 -18.8839  -0.2635  -4.1885
#>   -5.4933   0.3637  -1.2427  -3.9842 -11.0338  12.5833 -12.6536   5.4183
#>    1.8215  -0.0892   5.0791   2.1645   8.1993   6.6586  -0.2439   8.7508
#>   -0.1583  -5.9007   8.1124 -17.9357  14.6419   5.0039  -3.4683  19.0626
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]