Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  1.2628e+00  3.5611e-01 -6.8666e+00 -3.4319e+00  7.8210e+00  2.5531e+00
#>  -1.7936e+00 -2.6322e-02  1.1676e+01 -2.0140e+00  1.7951e+01 -8.9612e+00
#>   2.0307e+00 -6.9373e+00  2.3806e+00 -2.8814e+00 -5.5788e+00 -1.2474e+01
#>   6.5576e-01  2.9275e+00  1.4530e+00 -1.7458e+01 -3.2345e+00  8.1445e+00
#>   4.6836e+00 -2.3592e+00 -4.1845e+00 -1.9476e+00  4.8806e+00 -4.5854e+00
#>   2.1477e+00 -3.8433e+00 -2.4115e+00 -4.5822e+00  2.0408e+01 -7.2509e+00
#>  -3.4173e+00 -2.7148e+00  2.6390e+00 -9.8575e+00  1.3718e+01  5.8340e+00
#>   1.0947e+00 -2.7237e+00 -3.3465e+00  5.4362e+00  3.1465e+00 -6.0501e+00
#>   7.4035e-01 -2.5973e+00 -4.2992e-02  7.5065e+00  1.0951e+01 -1.4639e+01
#>   5.5402e+00  3.7983e+00 -6.2260e+00  2.7094e+00  1.1175e+00  9.6901e+00
#>  -2.4616e+00 -3.1432e+00 -1.8627e+00  1.5336e+00  5.1754e+00  1.2881e+00
#>  -2.8348e+00 -2.5643e+00  9.9450e-01 -7.2913e+00  6.1105e-01  2.8214e+00
#>   1.8378e-01  5.5263e+00  4.2086e+00 -1.3205e+01 -5.8842e+00  1.3809e+01
#>   2.8005e+00 -9.5219e+00  2.3340e+00  9.7669e+00 -6.1977e+00 -1.6745e+01
#>   5.5572e+00 -6.6566e-01 -7.4485e+00 -8.4955e+00  9.9944e+00  2.9348e+00
#>  -4.3304e+00 -2.0189e+00 -6.0923e+00 -1.1522e+01  2.2933e-01  8.0921e+00
#>   1.0569e+00  5.7142e+00  7.8189e-01  6.9288e+00 -1.5390e+00  9.6213e+00
#>  -2.7927e+00 -1.9296e-01  1.2135e+01 -9.7868e+00 -2.5043e+00 -8.6709e+00
#>   6.3063e+00  3.3212e+00 -7.7137e-01  5.8011e+00  8.4558e+00 -1.4863e+01
#>   6.4770e-01  3.7074e+00  4.8107e+00 -9.0366e-01 -6.9719e-02  6.8824e+00
#>  -4.8252e+00  6.7034e-01 -6.6215e+00 -4.6050e+00 -3.3858e+00 -1.5739e+01
#>   1.5201e+00  5.9362e+00  8.9834e+00  5.8181e-01  4.6662e+00 -4.0019e+00
#>  -8.6060e+00 -6.3775e+00 -8.5217e+00 -1.4043e+01 -4.1521e+00  1.7396e+00
#>  -2.7265e-01 -1.2351e+00 -1.0857e-01 -5.6955e+00 -5.1805e+00  5.4269e+00
#>   2.2925e+00 -2.8167e+00 -6.5598e+00 -1.0737e+00  3.4950e+00  6.4476e+00
#>   2.2678e+00 -1.5222e-01 -1.0924e+01 -2.7958e+00 -3.7777e+00  1.9120e+01
#>  -3.0299e+00  3.1364e+00  3.6060e+00  3.3095e-01 -1.8292e+00  1.5551e+01
#>  -2.9696e+00 -2.8456e+00 -4.9618e+00 -4.4105e+00 -1.2185e+01  1.2062e+00
#>   6.4899e-01 -1.0019e+01 -7.9954e+00  9.7914e+00  1.0946e+01 -4.6536e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]