Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.6804   0.3198  -3.1250   0.2243  -5.0378   0.9053   5.1499   3.9295
#>   -1.5412   2.6185   3.8735  10.8233  -7.2575   3.9903   2.9616 -11.0004
#>   -0.8324   3.6820   4.8396  -2.8559 -10.2451   7.0261   0.1846  -5.9631
#>   -1.8587   0.3448   1.2713  -1.9923   7.1789   1.1662 -15.1599  12.7248
#>   -0.7491  -0.6919  -1.8354  -0.9204  -2.1308  -6.0561   6.6900 -11.0719
#>   -1.3915  -6.1065   6.7478  -6.5479  -1.6748  -7.7132   5.1602   9.1457
#>    2.5645   4.3416   3.4206   7.0350   4.4948   6.0401   2.2420  14.8014
#>    6.7740   2.2798  -1.8575   5.3464  11.3414  10.3481  -0.1484  15.9856
#>    0.0164   1.9728  -1.2464   0.9035  -0.6432   0.0897   6.4972  -0.0160
#>    1.9533  -0.3126  -2.6008   3.3386   2.1282  -3.8620   2.1617   4.1248
#>   -2.0758  -1.4777  -6.2306  -6.0695 -12.4577   0.2245  -7.4696   6.6985
#>   -0.6363  -3.1828  -5.8696  -0.7340  -5.5470   3.6118   9.1955   2.6450
#>    0.1439   1.7644   3.8542  -1.3717   1.8496 -10.4253  10.5593   2.9960
#>   -0.8726  -5.2205   1.8624   0.7578  10.0075  -1.2039  -4.1519   3.2813
#>   -3.0187  -3.6543  -1.4344  -3.0463  -1.9706  -4.5387  -9.9128  -1.3637
#>    0.5075   0.3352   2.4439   3.7243   6.6171  -8.2706  -6.6569  14.8030
#>    0.0788  -0.1501   1.1661  -0.7348 -15.2179  -0.6778   3.4682  17.9118
#>    2.1126  -1.1918   3.6305   0.2900   1.5555  18.0215  -3.2440   2.0124
#>   -1.1620  -5.7314  -1.8169  -8.0019  18.2708   0.0592  -9.2557   6.8411
#>   -1.2273   0.1394   0.2607   3.2244   6.0700  10.2537 -11.2321  15.5584
#>    4.4647   2.0605   3.0482   5.8959   4.8856  -7.5090   4.8887  18.1799
#>    3.9718   4.2066   1.5404  -4.0519  -2.6310   4.0948  -3.0450   5.3950
#>   -1.6136  -2.2094   5.6694  -9.1711 -10.3634  15.6864   9.8723 -15.7283
#>    0.5781   2.6951  -1.8578   4.4991  -0.3424   5.6492  -2.2475   1.2804
#>    1.3224   2.4796   5.8623  -1.8299   0.5917  -1.2297   1.3216  12.6008
#>    1.3809   1.6056  -2.6549  -6.6950  -4.7727   4.9634  -1.0268  -0.0960
#>   -4.1735  -4.8005   4.0529   9.2008   5.9219   2.8249   0.0785   6.4773
#>    2.8553   2.0279   6.9988  -2.7081   0.6321   6.6543  -1.5262  20.2907
#>    0.4099   0.0190   3.9761  -3.4646   4.2631  -9.0327   6.6780   6.8232
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]