Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.5577  -3.7869   4.5619  -7.0086   5.0466  -0.6399   1.0420   8.2332
#>    1.6736   3.6708   3.5775  -0.6764 -14.8444  -5.9860  -5.7928  -6.0479
#>   -0.3214  -2.1522  14.1999  -4.3487   6.1771 -10.1876   5.8863   9.7922
#>   -1.5300  12.7959   6.4444   5.3143   1.2287   7.0612   1.7473   1.7623
#>    9.9825  -9.5529   2.7472  -8.1048  12.1062  10.9311   2.5231   6.9817
#>   -2.6439  -2.3094   1.1066   0.6519  -5.1818   2.3044   2.8626  -8.6735
#>   -2.5657   2.6427   4.4962   5.3619   3.0510 -11.1234  -2.7125  -4.1556
#>    2.1314   0.2602  -0.5079   6.6675 -16.3624  14.9238 -18.1124  -2.5639
#>    7.4014  -2.5980   5.0797  -6.3086   5.7751  -2.0884  -9.0401   8.8206
#>   -0.1731  -1.6658  -2.0267  -0.1967   8.1868  -5.5561  -1.7779  -0.0921
#>    1.4296   2.6082  -0.7112  -6.1889  -4.3773 -14.5997 -11.1680  -4.5829
#>    4.1829  -6.1091  -3.5288  -2.5529   0.2804  12.6463  -5.1163  -4.6424
#>   -3.1031   2.3517   3.4978  -7.0555  -1.2761   1.4035   9.7833  10.4224
#>   -2.3605   1.3615  -1.1777  -6.2259   4.6213  -2.4465  -5.7156 -17.7198
#>   -0.3360   1.7168  -7.1857   8.0249  -0.9650   4.1777   9.7310  16.6371
#>    0.6928   7.6094  12.2076  -2.5448  -4.1846   7.9447 -16.1554   6.2945
#>    4.4100   5.6280   0.4761   9.3591  -7.7606  12.6016  -7.3520   4.0268
#>   -0.9846  -2.8495   1.7051  -5.5949  -1.2146  -2.0920  -0.4950  -8.3723
#>   -0.8176   0.5871  -8.0229  14.2257   3.6532  -3.5893  -4.1223  -1.0662
#>   -1.1007  -5.7390  12.8937   2.4104   2.9647  -1.4996  -6.6574  -9.5316
#>   -0.4920  -3.6017   4.1992   5.2395  -3.1894  -0.6881  -6.3938   0.4575
#>   -1.4045  -0.4301  -5.0945  -6.7981  -0.0408  -2.3781  -5.2419  -5.5383
#>    5.0757  -7.8165  10.6776 -10.7716  -9.8275   2.8257 -15.2436   9.3712
#>    2.5515   2.2455   3.2905   3.4990  -9.4539   4.9406 -11.9686  -1.7730
#>    2.6546   1.6138 -11.5759   0.0902  -1.0130   9.0030  -9.4788   4.4822
#>    0.7726  -2.5527  -4.1381  -2.4196  -7.9439   4.2964   4.0163  -1.4218
#>   -1.7582  -0.8789  -1.8257   2.7105  -4.3723  -3.2165   1.3682   1.1448
#>    3.5013  -2.0940  -8.6615   8.7389  -7.1590   7.5455   8.6741   0.7132
#>    3.7305  -5.3569 -11.7343   7.1618  -1.9251   8.9230  -2.1528   1.9418
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]