Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.6526   3.6131   6.3987  -6.7598   2.4789   3.5769   9.2476   5.3831
#>    6.4707   1.7898  14.4142  -6.4775   1.4354  -3.3110   6.1961   1.3037
#>   -1.1831  -9.1780  -6.5980 -18.4578  -3.4791  -2.9013  15.7228   1.6399
#>    2.6873  -2.3160   1.2202  -9.5366 -17.0741  14.7228  -5.3331  14.9949
#>   -3.3653  -0.9571  -3.3238  -5.2555  10.8655  -7.9475  10.0201   1.8168
#>    1.6110   8.6932   3.5102   4.9112  -7.0386 -13.3749 -11.2774   7.6977
#>    0.9033   4.8650   8.7797  -9.6874  -3.1994  -1.9080  14.4467   0.0837
#>   -8.3314  -3.2868  -3.7988   8.9579  -7.1727   9.9202 -12.5183 -11.2527
#>   -4.7210   8.5335  -4.7455   2.3533   7.6767  -0.7945 -11.2753 -10.4479
#>   -2.9413   5.7432  10.2759  -2.2335   6.3019   3.1421   3.7312 -10.9435
#>    5.6632  -8.0966  -8.9269 -11.7933  15.8689   5.1767  19.3859   9.0110
#>   -2.9385  -4.2042  -8.8669   3.0404  22.8731  11.4588  -2.2838  13.1815
#>   -7.6728   3.7071   6.5102   3.2092 -10.1565   3.8693  -5.3316 -11.8196
#>   -4.1050   2.9151  -3.8793  11.4060  -0.2124   0.9337  -8.2135   9.6108
#>   -5.0544   7.5353  -8.6831   3.9182  -5.3565   4.8590 -11.2681   5.1881
#>   -0.1198   8.3727  -1.7476  -1.2516  -5.0379   8.3440 -15.0507  -3.5784
#>    2.7412   3.0480   1.6859   9.0003  -6.5768   6.2713  -5.9631   2.4205
#>   -0.3771   4.9757   7.8435  -4.4491   2.5529   3.7900   3.4531   3.2659
#>    3.6612   1.4240   5.5153   1.4355  -2.5264  -2.0226  -9.1845  -2.4176
#>    1.4472   0.1697  14.8696  -0.7285   4.4024  -6.0096   0.2251  -3.3158
#>    0.1136   4.8866  13.2016   2.8708  -2.6021  -7.5737   6.2828   4.6224
#>   -1.2486   6.1565   4.2042   1.4601  19.1954   7.5530   3.6181 -10.8277
#>    0.9584   0.3879   1.6221   3.0074   0.5880   7.8471 -10.4047  -7.4765
#>    0.0164  -8.9945  -5.2481   0.2712  13.5217 -15.1063  10.6621   1.1692
#>   10.4115   1.6245 -10.1239   2.6267   9.1080  -3.3815  -6.8400   1.6329
#>   -2.4998  -0.3651  12.9037  -4.1953  -1.8301 -17.6946 -13.6378 -20.4072
#>   -2.6346 -12.6759   8.5047   4.1149  14.7865  -5.7774  -0.1437 -11.5594
#>   -1.2320   2.7549  11.9157 -12.4004   5.6528 -16.5707   0.9608 -13.8402
#>   -1.2239  -3.1773  -7.5433   4.8358  16.2016   5.9481   7.6732   9.7114
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]