Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.4145   5.5688  10.7016  -3.1192   1.2499 -17.6659 -20.0628  -0.8251
#>    2.9367   5.4609  -3.2620   2.0470  -6.1936   7.6249  -7.1885 -12.3080
#>    4.0614  12.1508   0.9494   5.8593  -5.5665  -8.0219   7.9678   7.4910
#>   -1.9249  -8.0204  -0.0075   7.1593   1.4254   5.8177 -14.6217  -7.0497
#>    3.6245  -4.6125  10.5144  -0.3051   2.4750  -5.1382   0.2141   8.5682
#>   -1.0765  -1.6596   5.8572  -5.9381  -7.7522   0.1697  17.1412  -2.0474
#>   -1.4002  -1.7078   6.5085  -6.3037 -11.8774  13.8545   7.0373   4.8006
#>    3.9313   5.7888  14.1665   3.9775 -13.8756  -8.9889  -4.7473   0.9058
#>   -3.0140   3.7110  -7.3866   5.1254   1.3673  12.5774  -3.2686  11.1299
#>   -2.7905   3.1067  -2.0781  -2.4480 -19.6742   5.9980  -5.1098 -15.5711
#>   -2.9134  -6.8380   3.4924   1.4948 -21.2750 -12.3292   9.1422  16.5751
#>    0.5519   1.2082   8.6718 -11.3592 -10.4253  -2.0327  11.7073  -4.1452
#>    1.7975   8.1977  12.4363   4.1970  10.5434  21.2327   0.2263  12.9455
#>    0.4578  -3.8832  -0.7518  17.2868   8.6018   7.6407   3.1442 -14.5438
#>    3.7897   0.0331  -1.9716   1.8191   9.8711   1.7788  -4.0705   1.7388
#>    5.2460   1.1162   2.4008   8.2609  -5.5558  -3.1999   3.1694   4.9703
#>   -0.8927  -1.4633  17.7592  -4.0671   2.1570   2.4497  -5.1087  16.3435
#>    1.8366  -6.4254  -2.6213 -11.9010 -27.2605  -5.8572 -10.2660 -23.6821
#>    2.2845  -2.4047   6.3701  -1.2237 -15.4396  14.8779  11.9843 -15.2622
#>    2.0263  -7.0469  -6.8637   7.3828  -4.5341   8.9932   7.4999 -13.7392
#>    0.2931  -4.8527  -6.6090   0.0046 -12.8512   7.8685  11.1183  -9.2907
#>   -4.2884  -1.1467   5.2543  -3.5119   0.2666  -1.3846 -23.2015 -16.3591
#>    4.7075  -2.5575   0.8210  -4.8653  -9.0443  -4.9399   4.7120  -3.7274
#>    0.4407  -3.6878  -3.0652  -3.3628  -4.9718   7.5526  -1.3620   7.6958
#>    1.1118   1.1842 -14.1933  -3.9718  -2.2074   3.9380   9.5843   3.7659
#>   -4.4591  -2.5297   5.0816  -8.8131   3.8084 -12.7446   1.2334   2.5316
#>   -2.9046  -6.9842   6.8695  -1.9807   4.0315   3.4707  -1.4579  -3.9143
#>    5.6610  10.1397   6.1822   7.6502  -1.9683  -9.0705  11.9847  11.1279
#>   -0.5202   8.8840   1.2547   2.9261  15.0499  16.0686  -7.2172   8.9928
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]