Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.2008   8.0362   8.0438 -10.2365  -7.5039  -5.8277 -11.3786   0.8889
#>   -2.8027  -4.5929   6.5260  -2.7012  14.2028   6.2027   7.5092  11.6182
#>  -14.9018  -7.0173  -8.0541   1.3612  -7.1996  10.2693   2.7066   4.0325
#>   -6.0579   6.5659   3.6947   2.4840 -10.3163  -4.5439  -4.4479  14.2397
#>   -0.3582  -4.7328   3.9052   2.4329   8.4623  -2.9270  -6.1057  -2.1193
#>    1.3175   2.6547 -11.4162  -5.1062  10.0351   2.7809  -8.4992  -5.2414
#>   -0.3568  -6.8986   8.4818  23.3068  -7.4882  -9.9900  18.4099  -7.1733
#>   -1.6404  -2.6916 -12.2923   4.0393   3.4453   2.9788 -15.7198  11.1646
#>   -0.4085  -4.8255  -6.6488   8.2697  -1.3562  -5.6203   8.3475   4.4247
#>    3.8294  10.4443  -0.7792  -7.2803  37.5601  -3.9902   2.5648  -8.5389
#>   -1.0281  -1.0536   4.3534  -4.8996 -10.9025  -4.3243  -7.5950  -6.4083
#>   -6.8095  -1.9665   1.8115 -18.5928   0.1118  -8.4151 -13.1373 -20.6240
#>    4.8209   0.4500  -6.4886  -6.1824   0.7973   2.0227 -13.0788  -6.7763
#>   -4.9102  -6.3999  -3.8780   5.4073  12.8282  -3.8138   4.7809 -18.7810
#>   -4.5608  -7.9110  -5.3436   9.0292  -0.2386   9.7217   7.7691   7.0603
#>    1.9491   6.0570   7.2037   0.8726 -12.0847  -2.5239 -20.2427  -7.3190
#>   -1.4112   5.4442  10.2109   3.1184  -2.5156 -10.7956  -2.7735 -20.3585
#>    0.2651  -4.0488  13.1852   3.0206  -2.3714  -7.0532   3.6284  12.4263
#>    7.5909  -0.4438   3.8587 -11.9637  -8.2047 -14.8895  10.0308 -10.2199
#>   -4.8496  -2.6453   8.5338   3.7349  -1.0265  -0.0273  -7.0637  18.0140
#>    0.1778   0.7694  10.2850  17.7781  17.6345   8.2975  10.3746   9.0470
#>   -3.8490   5.3331   8.8510  12.5621  -0.9537  -3.7469  -1.5393  -3.4735
#>   -6.7547   3.0259  -3.7440  15.4211  -1.2271  17.1009  -0.5743  26.1006
#>   -1.5164  10.9053  -1.8808   3.5781  -0.4646  -4.8024  -9.0065 -20.6528
#>   -1.2312   0.1143  -8.5671 -19.4006  -5.4125  -2.7774 -13.9871  -4.9201
#>    7.0305   0.2959   0.7564  17.1239  -2.0991   0.7833  -1.8599 -10.2970
#>    4.7197  -0.0339   1.0978  -6.0604  -7.1830   2.2239 -23.7292  -8.2014
#>    2.1153   5.5529  -1.4594   6.7451  -0.2033  -1.0463  12.3946  10.5882
#>    0.5854   4.6848   2.8504  -4.2130   0.4479   4.8446  -8.0503 -18.0850
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]