Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.2911  -2.0476  -2.7091  -0.1869   1.8093   0.5798  -7.8445   6.3766
#>   -2.7845   0.2549  -1.5239 -12.8560   2.0056  -0.0264   1.9427  -6.8147
#>   -5.4788  -4.4567  -0.0605  -1.5042   1.9488   6.6688   5.7046   7.8602
#>    1.3384   1.6353   2.4839   5.2788  -7.9705   2.7634  -5.9020  -0.6126
#>   -1.5587  -9.6407  -6.0462   8.0912  -0.9639   3.6065  -6.9035  11.0758
#>   -6.9210   3.4651   3.8700   5.5157 -12.4818  -6.3578   1.6755   1.6328
#>    7.9514   8.4439  -2.2554  12.9740  -9.1023  12.2645  -6.8703  -5.9879
#>   -7.8715   2.2145   3.0013  12.0213  -3.6311   1.7916  -3.9982 -10.6722
#>   -2.8408   8.2937  -4.9334  10.4222  -8.0150  -2.6800  12.1471  -5.1737
#>   -0.6712  -7.2505  12.7973  -0.0911  -1.4906  -5.5917   2.6394   2.6947
#>    1.7836   4.1112  -2.1079  12.3851  10.2993 -16.5774   1.7089  -5.2634
#>    2.4519   9.8073   1.9331  -2.4255   6.8747  -3.5354  12.3868   5.6936
#>    0.5780  -0.7609  -1.4526  -0.4113   7.5699   2.5042  11.5368  -1.0985
#>   -1.1199  -5.2900  -1.2539  -6.3763   4.5898   4.0246   4.1523  -8.7309
#>   10.3200   4.6153  -5.2216   2.9315  -4.4359  10.3932   1.9464   5.1309
#>    6.4328   2.2914  12.9800  -3.9225  -2.5655   8.3224  -3.4053  -2.4349
#>   -3.7259  -4.6506   7.4969   2.5353  -1.5057  -2.9249   5.6022   3.9285
#>   -4.5245   0.0854  12.6328   1.8751   2.3845   7.1032 -14.5325  12.6659
#>    4.0906   7.0744   0.1601   6.1483  -6.7131   1.9149  -2.4562  -0.6433
#>    4.8296   4.7989  -2.9996  -1.6868   6.7002  -2.0048  -5.3640   8.1838
#>   -7.2162 -12.5938  10.6317   4.0927   8.8144  -3.8445   8.6486  -8.3884
#>   -0.8220  -5.9786  -0.8937   6.3761  -7.6120   0.4670  -7.5211  19.4641
#>    5.7180  -1.3197   4.2321   3.2871  -5.7875  -4.4620  -5.5851   6.2086
#>   -6.0277  -9.6928   6.4511  11.4390  -1.0929   0.9338   7.3301  13.8508
#>   -0.2164   1.3135  -6.9405  -4.9937  16.1941  -0.5604  17.0094  -7.5332
#>    5.6022  -0.7488   9.0451 -16.0573  -7.3598   9.8672   1.0684  19.6441
#>   -1.2178  -0.9345  -3.0523   4.1403  -2.4365 -11.0442   0.4281   0.9348
#>    0.7664   5.3909 -13.9889   1.3783  18.4085   7.4423   0.9132   2.6257
#>   -1.9571 -12.2830   5.2131   8.5454 -16.8639  -7.9428  -6.5896  -2.0043
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]