Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.4485   1.7888   2.4251   1.8434 -11.4239   6.8004  15.9167   2.4684
#>   -3.2593  -0.5379  -5.1588  15.1633  -4.5599   8.5072   9.6375  -1.8514
#>   -3.0349  -5.1850   0.0920   7.4944  13.1411  13.0787   4.2692  -0.5682
#>   -3.7088  -3.6766   1.7920   2.7905 -14.4378 -10.3739 -19.9470  -6.9395
#>   -8.3828   5.8735   5.8570   7.1062  13.5472   3.8366  -5.3569   7.0379
#>    7.1995  -9.7300  -6.7243 -15.1956  -6.2423  -4.0297  -3.1453  29.3301
#>   -2.5245  -7.6809   5.3269 -10.2952   3.1206   0.2951  -8.7517  -6.5333
#>   -5.6970   0.6418   2.7224  -5.3661   0.1828   3.3323  10.8281  -5.9192
#>   -2.0463   0.8474  13.9709   9.7033  -8.7181  17.2352   7.9381   2.0440
#>   -1.9536   2.4189   7.7938  11.8456  -4.2269 -16.7862  -8.4592 -16.6310
#>   -1.0536  -0.1645  -3.5228   7.0608  -4.3841   5.6575  -3.1235   1.1405
#>   -3.5782  -1.7072  10.1282  -2.5500  -7.8208  -4.5697  -8.4298  -2.7954
#>   -1.0706  -1.3423   4.7980   0.4580   0.3945 -11.6560   3.6451   5.2711
#>    5.3052  11.0945  13.5154   5.4370   6.3778   4.5791  -1.9317  -1.1053
#>    0.6126  -1.6661   6.0271   2.3011  -7.2842 -21.4366  -3.7534   6.6814
#>   -1.4664  -8.4369  -2.4456  -1.1089  12.1097   5.9290   4.1333   4.7083
#>   -0.3687  -0.3356   7.4410   5.6368  -8.4635 -11.3699  10.7925 -18.2353
#>    4.0059  -5.3369   1.5808  11.5410  -9.4336 -19.4638 -11.1769  19.3420
#>   -0.7686   0.6528 -11.3400 -12.9878   2.6498  12.1897   2.0382   0.7519
#>   -1.9452   0.2010   7.0690  23.5988   6.6439   3.2060  -2.1196   2.5404
#>   -8.6237   5.0103  -5.7568   2.8004   5.6692  -4.0955  -2.0843   0.7516
#>   -4.3006   4.3138  -6.2518  -1.2721 -12.9883   6.7130   1.3896  -7.2358
#>   -1.3665   7.3010  -1.1954  10.3300  -2.7820   3.4417   1.1584   9.0786
#>    0.8610   0.2953  -8.0666 -10.3192   0.3392   0.2455   1.8636   1.9973
#>   -3.5772  -3.7463   6.6032  -4.8441   2.6252  -7.0597  -0.4371 -21.2297
#>    0.5393   4.1055  -0.6048 -15.3675  -0.7843  -4.3137  -9.8649  12.8255
#>   -3.9050  -7.8905  10.6537   2.9378  12.1830   5.2406  -2.0369 -18.5555
#>    2.9338  -1.1910  -9.0209  -0.7278   3.5073   4.9269   1.5702   9.0760
#>   -2.9244   1.9959   2.2903   4.4672  10.9440  -7.5618 -10.3335  -3.4616
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]