Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  2.2544e+00 -7.2308e+00  2.2101e+00 -1.1791e+01  8.1589e+00  1.4687e+00
#>   2.5473e+00  1.4431e+00 -8.5129e+00  5.4726e+00 -6.8853e+00  8.1588e+00
#>   2.9301e+00 -1.5032e+00 -9.2471e+00  2.7909e+00  4.7965e+00  2.0328e+00
#>   1.3508e+00 -1.0431e+00 -4.4070e+00 -6.7403e+00  1.2079e+00  4.9491e+00
#>   4.9302e-01  2.6538e+00 -2.1958e-03 -2.4323e+00 -7.0839e+00  1.5468e+01
#>  -2.2022e+00 -1.1352e+00  2.7207e+00 -6.6491e+00 -5.3153e+00 -5.6771e+00
#>  -1.7744e+00  8.6555e-01  2.8181e+00 -4.6044e+00 -1.3514e+01 -8.8215e+00
#>  -1.0748e+00 -4.5665e+00 -3.5756e-01  2.7175e+00  7.8508e+00 -1.8316e+01
#>   1.6694e+00 -5.7435e+00  2.4992e+00 -3.7312e+00  6.8539e+00  1.1887e+01
#>   9.2079e-01 -1.8633e+00  5.0631e+00  2.0022e+00  9.0192e-01 -1.6318e+01
#>  -5.0439e+00 -4.7050e+00  6.8121e+00 -8.3224e+00  9.4566e+00 -2.1882e+01
#>   2.4437e-03 -5.4226e-01  4.4266e+00  2.9370e+00 -9.6463e+00 -2.8288e+00
#>  -3.4274e+00 -1.9366e+00 -5.0681e+00 -7.5870e+00  3.2299e+00  1.1026e+01
#>   6.9224e-01  2.2033e+00  2.9232e+00  2.1331e+00 -4.2311e+00  9.4411e+00
#>   6.2600e-01  3.9985e+00 -3.4583e+00 -1.2496e+00 -5.9442e+00  8.4790e+00
#>  -5.0238e+00 -3.3297e+00  2.8342e+00  1.3430e+01  8.8786e-01 -3.9483e+00
#>  -1.2940e+00 -4.5638e+00  2.1123e+00  4.6221e+00 -1.0912e+01 -9.8534e-01
#>   2.9330e-01 -1.7468e+00  5.0964e+00 -1.6079e+00 -1.5592e-01 -8.8139e+00
#>  -2.5941e+00  1.0287e+01  5.3087e-01  1.2998e+01 -1.0469e+01 -1.5923e+00
#>  -7.9610e+00  1.8680e+00  3.8759e+00 -2.4214e+01 -2.0988e-01 -1.3578e+01
#>   7.7553e-01 -1.8486e+00  3.9549e+00 -5.1785e+00  3.0503e+00 -1.3511e+01
#>   4.7304e+00 -6.7654e+00  4.6200e-01 -4.9312e+00 -3.0531e+00  1.0827e+01
#>  -1.6431e+00 -6.2343e+00  5.3247e+00  4.9266e+00  6.9432e-01 -1.5623e+00
#>  -3.7325e+00  2.1177e+00 -4.8250e+00 -7.0884e-01  7.5816e+00 -2.5354e-01
#>   4.5211e+00  2.0093e+00 -3.8531e+00  7.6763e+00 -1.2462e+01 -9.5679e-01
#>  -4.2582e+00  1.1997e+00  3.3589e+00 -6.3600e+00  6.1465e+00 -2.6464e+01
#>   1.1078e+00 -1.5521e+00  1.0955e+01  4.5292e+00 -3.7499e+00  1.6553e+01
#>  -1.2272e+00 -8.9431e-01  4.1852e+00  2.7913e-01  1.2577e+01  2.4073e+00
#>  -2.1598e+00  2.3916e-01  5.8247e+00 -4.4659e+00 -9.3153e+00 -8.9744e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]