Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.1117   8.1452   2.2027  -1.4898   4.7818  -3.8238  -7.7560   1.1701
#>   -5.4430  10.4224   1.9065  -1.8287  -0.4101  12.4602   0.9283   3.5371
#>   -1.4395  -0.3406   7.2593  -2.4053  -6.5444   1.9493 -13.8817  -3.6689
#>   -2.8917  -5.5626  -3.8516  -7.5155  -3.6382  -0.7414 -11.4640  -9.9395
#>   -3.8157   1.2681  11.9069  -3.5489  12.5200  -2.3802   2.4375  20.1446
#>    8.5635  -4.1071   0.1562  -1.2706  -6.3687 -14.7124 -12.8036   0.5650
#>    0.0473   8.6760   1.9900 -11.9703 -16.8728   4.9995  -5.1642  -8.6304
#>    2.5657   1.0795   7.9579 -10.7930  -3.0856  16.4950  15.6979  -4.7961
#>    6.8706  -2.4624  -0.8018  -9.5757   3.0610  -8.6981 -13.7777   6.0757
#>   -2.7411   0.1495  -5.5464  -5.7535  -4.2466  -2.9632 -16.3137 -11.5539
#>    0.3719  -4.3285  10.9740   4.4416  -0.8697   1.0343 -13.9108  -6.9198
#>    4.0340  -9.3014  -1.5914  -2.9583  -5.9644   0.7987  -4.1117  10.3757
#>   -4.5476   6.2829  -2.7022 -16.3538 -12.0263   4.1056   6.9692  18.5467
#>   -6.7253   3.1063  -3.6227  -9.9085   2.5019 -14.3513  -3.2368 -10.0289
#>   -3.4341   0.7228   3.1057  10.0190   6.3801   6.5556  -6.6094 -10.5100
#>    1.5182  -4.1525  -5.6908   1.0407   7.4344   3.8868  -6.1672   7.3084
#>   -4.1859   2.6365  -2.0610 -12.8335  13.7657   5.5780   2.3923   6.6115
#>   -2.7107  -3.8254  -4.9595   3.7181 -11.1801   4.6870  -4.7271  -2.6277
#>   -0.2876  -1.9815  -0.1674  12.1714   7.9775  16.6708 -10.3321 -17.8466
#>   -4.6066  -0.5514   9.4385   2.7748   6.9368   4.4314 -18.1496   0.2602
#>   -9.8444   4.1539  -4.3505  -6.0942  -7.3545  12.2350   8.5204  11.9469
#>    4.2188  -4.0418   1.9582   3.5649   0.8452   3.2816 -14.4975 -12.7372
#>    2.2603   6.1410   3.8172   8.1759  -7.2474 -10.9651  10.3411   0.7411
#>    0.6104   2.4115  -3.5606  10.1179 -10.0837  12.7100   8.0088  -3.1451
#>   -5.8544   1.0168  -2.6044  12.2382   7.5363  20.0249  14.1234 -10.6745
#>   -7.4211   0.9819  -4.1309   0.9935  -8.3107  -3.7127  11.4608 -12.3638
#>   -1.4983  -3.7974  -4.6510  -4.2339  -0.9152  -2.0802   8.3283  -9.7986
#>    5.3456   4.1936  -9.2523   4.8967  16.9642 -23.5798  -5.9742  -5.4437
#>   -6.4837   2.4275 -14.3383 -13.8759 -16.0560  -1.0813  10.4663   0.0255
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]