Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.8183  -4.0091   7.5908  -8.8366  -3.9336  -5.9552 -10.3397  -8.3442
#>   -2.3644   4.2853   1.3035   4.3586  -3.8093   4.4566   6.9960   1.0182
#>   -1.0831  -5.6088  -4.6377   0.5175  -0.5008   3.8651  -4.8403   1.0849
#>    0.6207  -1.6930  -8.5513  -4.8601  -3.2698   9.6643   2.7487 -17.5616
#>    1.3037  -2.1879   6.4595   5.3316   3.5238  -8.1231   8.5144  -0.2737
#>    7.2832  -9.8637   3.1619   5.8023  18.6716   9.2022  -2.8994  -9.8570
#>   -4.9296  -6.7961 -12.5664 -12.9816   7.1625  12.2876  -1.0316   1.9645
#>    0.1873   2.2532   2.8314   1.8680   5.3981  -7.9527 -12.5607 -11.5363
#>   -3.5720  -4.8391 -12.2222   5.6619  16.7870  16.8347   2.1536 -11.9325
#>    5.2372  -2.6115  -7.1331  -7.6209  -5.7240  -8.2883 -14.9979  -5.3490
#>    9.1595  -4.5757   1.5100   2.2565  -3.9191   2.0654  -8.0780  18.9587
#>   -1.6011   0.2140  -8.8690  -3.5071  -1.6136   3.8644   2.7765  -2.5847
#>   -9.6821   1.5714  -3.9553  -5.1375 -11.3295 -14.3700   5.7197   5.8562
#>    1.8099   7.7631  16.5610   2.7301  -0.5285 -12.4036   5.7326  -5.0301
#>   -6.6074  -2.7700 -12.1164  -4.2170  -2.7694  12.5057  -1.9319  -6.1886
#>   -5.3295   1.4898   3.1598  -3.3522  -5.6124  -2.9899   1.0113  15.2883
#>   -4.0442  -0.8110   9.5293   2.4324   7.1734  -0.1759 -18.2647   1.8818
#>   -3.0117  -3.9929  -4.7560  -5.0720  -8.5505   7.7403 -12.2135   2.1533
#>   -1.2248   2.6398  -2.0628  -2.9657   8.4082 -10.5087 -21.5654   1.8400
#>    9.2198  -1.6085  -0.0324 -10.4586  -7.1824   2.2584 -21.1343   3.2802
#>    3.7605   2.1049  11.3444 -10.0926 -13.1832  -9.8548   5.6461  18.0741
#>   -5.2073   3.8096  -9.7327  -4.3207 -12.4119  -7.7670  -0.8549   7.6470
#>   -3.6829  -2.6096   0.2661   1.5446  -7.7179  -9.2700 -18.1453   6.7463
#>   -3.5221  -1.5174   1.3659   2.2683  10.5366   5.8321  14.0242  16.6539
#>    0.4591  -7.3079   5.0299  -3.5904 -12.5074   6.6956  16.4906 -12.1575
#>    1.6990   7.3750   0.4539  15.9428   2.3456   0.9000   0.5703   4.8245
#>   -1.3166   4.8601  12.4715   6.3185  -8.5397 -12.0288 -12.7744  -7.5450
#>    1.5522  -1.9246  -3.6187  -4.9163  -8.4037  -1.4302  -1.3346  -7.1371
#>    2.9437   2.4392   7.3581   7.5175   2.9157   8.7794  13.0913  14.9333
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]