Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.7380   3.9895  -4.0476  10.9591  -3.0559   6.3545  12.9287  12.9059
#>   -0.6202  -6.1010  -3.5843   9.3319   1.2021   2.4147  -7.7737   3.0318
#>   -2.7987   9.9600   5.2202   0.8242 -10.5616  -5.0504   0.6207   4.4569
#>    2.6162   4.8684   0.9581  -6.9338  -7.2811   2.7884   4.5510  11.3231
#>    3.2453  -1.3170   4.3560 -14.3319  -3.1322  -5.3298   8.6476   2.0466
#>   -5.4602  -4.2551  10.2235   2.6021  -1.5335   0.1714  18.3994   3.1012
#>    1.1588   0.2036  -3.6316  -4.4676   6.6804 -19.4197   2.3392 -16.0287
#>    5.3174   1.7542 -10.8486  -7.9093  -8.4660   4.5294   8.6420  -4.2456
#>   -0.9184  -4.3103  16.9428   7.4543   6.1361 -12.3612 -10.4398   6.2973
#>   -3.1860  -7.9666  -6.4974   3.6214   5.4914   9.7958   5.8853  -0.2640
#>    4.0641   4.1614   9.1128  -3.8951 -11.4138  -0.7903  14.0450  -3.6020
#>    0.1999   5.2374  -0.4535   0.4291   6.2682   2.7567  -1.3153   2.2066
#>    3.8215  -7.4325  -8.3005  17.5584  11.4484 -24.1932 -16.5323  -4.0913
#>   -2.9407   1.5935  -7.4040   9.7730  17.3535  -1.4901  -0.5789 -15.2660
#>    6.4926  -8.3487   2.1912   3.4743   0.8137   2.4690   6.9254  -6.3673
#>    3.0563   0.2841   6.5403  16.4606  -5.8343 -17.8722 -15.2612  -8.2300
#>    2.5610  -3.6367  -3.7733 -18.3247 -14.2392  -4.5656   2.5168   4.1630
#>    0.0894  -3.3810  -4.1926   4.4384   3.2692   2.5261  -1.6244  -0.5583
#>    5.2644   3.3495  -4.4041  -3.8256   9.2544   1.0242   0.4003  -6.8092
#>    4.8423  -5.1308  10.0070  -9.0973  -8.6670  -7.1204  17.3877   5.1604
#>    3.2052  -6.4958  11.0656   3.0746  -1.9303 -24.5129   3.0687  -3.4868
#>   -6.2484  -1.1542  -0.2524  -8.8870  -1.6863   1.7883  22.4507  -9.4810
#>   -2.4483   1.7229  -4.7450  -8.4319 -12.1097  11.3131   3.8290   6.7970
#>    5.0990   1.8410  -3.8714   4.4130  -7.1398   0.6732  17.2268  17.0242
#>    0.6487  -6.0102   5.3594  -5.4161  -4.6393   0.2900  -4.3699   4.3668
#>   -1.2148   4.1201   0.5574   2.8478  15.9613  -1.6879 -16.4351  -2.1517
#>    3.4479   4.5073  -2.2534  -6.5376  -7.6808 -15.6894 -12.0074 -16.2355
#>    0.2986  -2.3092   6.8490  -4.4263   3.9286  -4.9904  -8.6145  -5.0756
#>    3.6135   3.3272   6.5489  -2.6563  -5.2697 -13.7110 -12.0389 -11.4454
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]