Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.4465   7.6873  -3.2671   8.0118  12.4041   7.6183   4.6236   2.9160
#>   -2.2182  -7.2104   8.7581  -0.3859 -14.3607   8.3702  13.3407   3.5534
#>   -2.0757  -4.7255  -4.9660   0.6556   1.3758   3.6533   5.5488  11.7555
#>   -3.3037  -7.1065  -7.4317 -15.5615   2.7397  -7.1919   4.7661  -0.6787
#>   -4.7081  -8.3326  -0.7980   5.4690  -9.2371  -7.8960  14.7032  -6.8084
#>   -5.7167  -8.0597  -3.6601  10.0516   3.0187  -7.4824   7.2997   8.6951
#>   -5.2359  -1.3896   6.4196  -0.5146   2.6504 -12.3637  13.0934  -0.2743
#>   -2.1038  -1.8017  -1.9794   2.7467  -0.0941   2.0927   2.0520   6.7717
#>    0.8352   1.1881  -2.3080   7.6915   6.3264   7.0261  -5.3387  -9.0870
#>    0.4483  -0.4078   5.9266   5.4107   3.1694   4.2142  -3.7169  -9.6409
#>    0.8628   7.3771   6.8784  -3.4761 -10.3092   0.3406   3.0047  -6.5315
#>   -6.4262   4.6971   6.1729 -13.8627   5.1683 -11.1583   0.2335   1.3256
#>    0.3815  -2.3481   2.3680   9.1282  -2.8069  -5.6675  10.3015  -0.3994
#>   -1.9829  10.3464  -1.6707  -0.8120   8.5890  10.6347   1.3640  -5.8963
#>    6.1846   1.1235   4.6379  -1.2615 -18.4293  -9.2451   3.2897  -2.3618
#>   -3.7052  -1.2356   0.6478  -5.1747  -5.4434 -13.4410  -6.9681 -11.7596
#>   -2.0809   7.4054  14.1599   4.9573  -3.7688  12.9795  13.3000  -2.2505
#>    5.7953   2.2524   6.1586  -1.2398  10.4505  -6.3530  -0.0925  -7.9346
#>   -2.5992  -1.9101  -8.4936   1.1287  14.9367  -9.7904  -4.2768   8.8569
#>    5.7732   1.8283   4.1356   0.4701   4.2287 -10.4894   1.4449  -4.5254
#>   -2.3381  -5.4648  10.1083   0.1774  -4.6335   5.2231  -4.8461 -11.6607
#>    8.7111   6.3297  -0.2476  -4.2847 -12.6655  -8.5906  -1.0877   4.7014
#>    2.5295  -1.5338   0.5270  -0.5524  -5.9085   0.6373  -2.1854  -0.9114
#>    9.4449   6.5419   2.1156  19.0797  12.5033   4.0806  -7.7474   7.1470
#>    7.1392   3.2362  -1.3489   4.0942  -6.1418  -9.8016   5.3655 -13.8467
#>    8.4695   3.7066  -8.4811 -13.2011  10.5309  -0.1845  -2.2434   8.2718
#>    3.0360  -2.4864  -1.2526  -8.7214 -11.0850  -2.0828  15.8003   3.6940
#>    1.6246   0.4645  -2.7891   6.0401  -8.5571   0.2560  -8.0566  21.1448
#>    3.6232   3.3458   9.6160 -10.6172   1.0068   7.0524  -0.4916   1.2295
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]