Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  3.0349e+00 -1.2427e+00  3.7460e+00  8.8542e+00  3.5097e+00  8.9968e+00
#>   5.7577e+00  2.0644e+00 -2.3230e+00  1.1504e+00  1.1061e+01  4.5770e+00
#>   2.5806e+00  9.9648e+00 -4.5547e-02  9.0642e+00 -2.5770e+00 -3.6205e+00
#>  -2.9231e+00 -4.1650e+00 -1.5012e-02 -1.6856e+00  4.2701e+00 -5.4150e-01
#>   6.4414e+00 -1.2833e+00 -2.3893e+00 -8.5821e+00  2.1411e+00 -1.1793e+01
#>  -5.2150e+00 -1.8446e+00  8.1524e+00 -2.1871e+00 -4.4879e+00  5.1836e+00
#>  -6.0587e+00 -3.3244e+00 -3.5288e+00  1.4500e+00 -1.0300e+01 -9.6306e+00
#>  -2.1192e+00  3.8064e+00  6.0070e+00  1.7194e+01  5.8220e+00  5.2388e+00
#>   7.6912e+00  8.1348e+00  1.5026e+01 -7.1731e-01  4.8955e+00  3.5382e+00
#>   6.3000e+00  1.3833e+01  7.1301e+00 -3.9562e+00  3.0336e+00 -2.9670e+00
#>  -3.2329e+00 -5.8016e+00 -3.7046e+00  3.2949e+00 -1.2781e+01  2.4868e+00
#>  -1.4812e+00 -1.9179e+00  4.1436e-01 -4.4655e+00  1.1389e+01  8.8736e+00
#>  -3.4962e+00 -3.4094e+00  1.6074e+00 -6.0602e+00  3.4658e+00 -3.7124e+00
#>  -6.5630e+00 -1.5277e+00 -1.2359e+01 -3.3395e+00 -9.9788e+00  6.8686e+00
#>   4.2590e+00  4.5544e+00  7.4252e+00 -1.5773e+01  1.3826e+01 -1.3302e+01
#>  -3.5418e+00  6.7610e+00 -5.1601e+00 -1.4885e+01 -7.9599e+00 -1.2350e+01
#>  -9.4168e+00  2.1932e+00 -1.1168e+01 -1.0632e+00  4.6646e+00 -6.0739e+00
#>  -4.2469e+00  5.8303e-01 -1.3306e+01 -1.3695e+00  9.7111e+00 -1.0233e+01
#>  -2.8971e+00  7.3189e+00 -1.2549e+00 -7.4666e+00 -8.0992e+00  7.1139e+00
#>  -1.7035e+00  2.2575e+00  4.8064e+00 -1.3832e+01 -5.4856e+00 -1.3698e+01
#>  -1.7819e+00 -7.3177e+00  4.7782e+00 -1.0043e+01 -1.7944e+00 -3.1316e+00
#>   1.0858e+00  5.0453e+00  3.4424e-01 -1.6221e+00 -1.3935e+01  6.6547e+00
#>  -2.3185e+00  4.8791e+00 -4.7317e+00 -8.3252e+00 -6.7637e+00 -1.2191e+01
#>  -1.5065e+00 -1.3479e+00 -2.1904e+00 -6.1942e+00  9.5311e+00  9.2322e+00
#>   6.7960e+00  4.3213e+00  2.1988e+00 -8.9987e+00  8.5879e+00 -1.0133e+01
#>  -2.4679e+00  9.5855e+00 -1.1417e+01  1.8138e+01 -5.0434e+00  8.3484e-01
#>   1.3106e+00 -9.0750e+00 -1.0315e+01 -7.8525e-01  2.1026e+00  1.6824e+01
#>   1.0736e+00  4.7345e+00 -4.3392e+00 -6.1313e+00 -1.6675e+00  8.8436e+00
#>   6.2888e+00 -9.8332e+00  4.0159e+00  7.2451e+00 -1.6966e+00  1.7295e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]