Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.7634  -3.0865  -3.9332  -1.0014 -10.8604   8.3262 -15.9225 -12.4283
#>   -0.8729   6.0928  -1.3826  -4.3972  -3.0109   1.7312   2.3730  10.8430
#>   -2.9893  -0.3400  -9.7434   8.7426   2.9440  -1.9658   5.7622  -5.5502
#>    2.7335  -0.1046  -1.0515   3.1042   4.9814  -5.5208  18.1499  -1.0815
#>    2.2095  -7.7119   6.9504  -1.3688  -3.6390   4.3801  18.1701  -0.8270
#>    7.5840  -5.6694   0.0136  18.0740 -15.4811  -8.5871   5.4142   9.8608
#>   -1.6659  -0.7075  -7.6702  -3.5819  20.5776 -19.7057 -10.9102   8.2797
#>   -2.3812   6.4584  -1.8676   9.6716  -7.9915   1.6581  -4.7643   9.5507
#>    4.8334  -3.9139  -1.1194   8.1255   7.4955  14.3438  -4.4706  23.8492
#>    4.2900   4.8460 -10.0364   2.8087   5.4860  -0.4546   5.4268   0.7695
#>    1.4682   4.9310  -3.6171   2.2928  17.4437  -9.6582   4.5070  14.2377
#>   -0.7156  -2.1224   1.4110   3.0136 -15.3208   3.0341  13.8183  -7.7959
#>    2.1724  -5.3569  -2.2901   9.9570   3.5845  -9.1414  13.4320  -0.7019
#>   -0.1153   1.7543   3.0128  -1.8303  -1.5620   7.9682  -3.0859  25.9585
#>    3.1385   1.5985   9.4961  -4.3782 -13.9594  -8.3738  -9.5876  -2.9996
#>   -3.3489   1.1904   7.3337  -1.7613  -0.3940  -3.3136   0.5917   2.2328
#>   -6.6901   2.2478   1.0087  18.7107  -2.2772   4.0675  14.1052 -19.9251
#>    3.3818  -1.5137   7.4360   1.5683  -2.3790  -7.3046  14.9989   8.6297
#>   -8.1111   7.9415   7.9029   3.1978  -0.0786   4.1634   4.4335  -3.5497
#>    5.8954 -10.9257  -3.7954  11.9793  -7.7008  -7.5447  -1.8591  -3.5649
#>   -4.7780   5.8191 -11.1030  -4.9341  15.4340   4.3391  -8.7953  -1.7144
#>    2.4830   2.5207   9.2798   3.2010 -13.5786  -7.7594   6.2026  -6.0682
#>   -4.1113   6.2945  -8.3384  -6.7035   1.2069   3.6634   9.2165  -5.9552
#>   -7.5866   2.9308  -0.6159   2.9200   7.6388  -0.8991   3.4736   8.2256
#>   -4.2766  -0.3230  -2.1312  -6.2523  -8.8116   9.3325   1.1427   1.4430
#>    0.6689   1.5277  -9.7865  -6.2187   9.6553  -2.3570  -4.2839   1.6528
#>    0.0865  -2.7242  -4.9350   8.7851   3.4728  -0.2679  -2.8457   1.9261
#>   -1.8675  -1.9790 -11.5441  -1.9758  -4.3662  -6.9569   9.1407  -4.5954
#>   -3.3507   2.2899  -3.2968   2.3240  -3.1399  -0.4045  10.4911 -18.9874
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]