Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.1787   6.9023  -3.1384  -4.3312  -7.1523  -1.0225   8.3214  16.3972
#>   -4.4528  -1.0601   5.4955   6.3182  -5.7963   8.5683   6.7839  -4.5319
#>    1.2202  -3.9776  -1.3580  15.6418  10.8196  -5.0422   7.4489  -7.9365
#>    3.7969  -0.0437   3.1206   4.7424   6.6874  -5.3708   5.9958   9.1318
#>    0.0277  -1.2331  -0.0817  -1.6902  -1.6209  -1.9339  11.1238  -5.8560
#>    4.4497   1.4401  -4.7732   3.4570  13.1079  16.5528  -8.9384  -8.6528
#>   -3.4834  -5.7379   0.7184  17.2705   0.7623  -9.6190  13.8313 -15.1488
#>    1.2531   2.5085  12.7285   5.9908  17.6984  -3.8514   0.8898   7.9167
#>    0.4702   4.6607   6.9925  -5.4472  -8.1857   4.1019   4.4583  -0.2160
#>   -0.3774   2.7602   4.1346   4.7844  -1.2787  23.5028   5.0293   5.7949
#>   -7.0247  -5.3709   0.0673   2.6832 -10.0622   3.7660  -6.1636   6.1608
#>   -2.9971  -9.0920   8.2944 -14.5222 -19.4161 -12.4229  -5.5264  -5.3689
#>   -2.5177   1.4183   1.9275   8.8331 -13.1866   0.4426  -6.5216  11.6626
#>   -2.4251  -0.3901   8.3384   2.1803  -3.2199  13.0667  10.6630 -10.6396
#>   -0.2390   0.5157   4.3593   1.3993  11.8954  16.8480  -0.1402   9.8761
#>   -1.6624  -0.7562 -10.9032   7.3828  -3.7992   0.6078   2.7011  -1.9912
#>    0.4760   2.0727   5.4966   5.9981  -7.6896  -1.5823  -7.0156  -8.0101
#>   -2.3495   3.4528   0.1983   6.3584  13.2148  10.7317  17.5426   1.2031
#>   -2.4843  -3.8109  -3.6812  -3.0940   4.1315  -9.5185 -13.3583  -4.1301
#>    0.2393  -0.8208  -0.6572  15.8100  -3.8040   9.7934   1.9737   2.0162
#>   -1.1443   0.7523  -1.7244   9.2459   1.3714  -4.4704   1.5681  -4.4048
#>    0.0609  -1.4700  -6.4381  -3.6682  -3.5384 -10.9639 -10.7834   9.5648
#>   -4.8798  -0.9155  -5.3305  -2.8545   9.3076  -3.3131   3.3840 -10.5482
#>    5.5451  -7.7442  -1.3815   8.8951   6.0258  -2.9829  -4.6527  20.8493
#>   -0.2900  -1.0870  -3.9375  -6.5995  13.5239  -1.0754  -9.6536  -0.1406
#>   -1.3878   6.9486   5.9432   8.1547   6.1474   7.0463   0.1618   7.3388
#>   -0.9334  -4.5702   4.9294   9.0006  15.8651  -3.6990   7.5337  14.9992
#>   -0.7846   5.2305  -1.2900  -2.6154  10.5455   5.0067 -11.1756  12.0065
#>   -4.2476  -0.3348   4.4328   7.3309  -4.1282  -2.7860   4.5874  -8.9466
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]