Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -6.8232e-01  3.0315e+00 -1.9357e+00  1.5908e+00 -9.5138e+00  2.1440e+00
#>  -3.8675e+00  2.3526e+00  3.3989e+00 -6.7166e+00 -5.4525e+00  6.5903e+00
#>  -7.6534e+00  4.7657e+00  9.4863e+00  5.4834e+00  1.3524e+00 -1.4182e+01
#>   1.9051e+00  3.6293e+00  7.7896e+00  5.3870e+00  3.2478e-01 -7.3267e-01
#>  -2.8875e+00  5.5560e+00 -3.3272e+00  7.1891e+00  1.6457e+01 -2.1884e+01
#>  -3.2349e+00 -1.1316e+00 -1.1709e+00 -1.2861e+01  2.6259e+00 -1.5505e+00
#>  -5.0367e+00 -2.8438e+00 -3.0446e+00  2.1348e+00  8.0343e+00 -4.4652e+00
#>   3.3915e+00 -6.4505e+00 -6.9754e+00  1.2959e+01  3.4273e-04  1.2329e+00
#>   5.5455e+00 -5.5384e+00 -2.9448e+00 -1.2061e+01 -1.5875e+00  9.2337e+00
#>  -8.8703e-01  4.5879e+00  5.3296e+00 -4.8853e+00  5.1243e+00  6.4335e+00
#>   5.3747e+00 -6.3037e+00  7.9219e+00  1.1824e+01 -1.4392e+01  1.0144e+00
#>   1.6035e+00 -1.8362e+00 -5.6030e+00  3.4779e-01 -6.1093e+00 -2.1134e+01
#>   6.1772e-01 -2.4060e+00  9.1966e+00 -3.6593e+00 -8.2454e+00  3.6497e+00
#>   7.2019e+00 -6.2026e+00 -3.6167e+00 -6.3086e+00 -3.4853e+00 -2.6391e-01
#>   1.5775e+00 -5.7493e+00 -1.5701e+01  7.0557e+00 -9.7730e-01 -1.9120e+00
#>  -2.6932e+00 -6.6438e+00  2.1556e+00 -9.2254e+00  3.2108e-01  4.8204e+00
#>   2.6429e+00 -3.7126e+00 -8.1428e+00  1.5385e+01 -1.0685e+01  9.1858e+00
#>   6.3621e+00 -1.0269e+01 -2.7927e+00  2.1630e-01 -2.4322e+00  1.9797e+01
#>   2.9713e+00  4.3156e+00 -1.6753e+00 -2.3846e-01  1.5180e+00  1.6064e+00
#>  -1.1022e+00  7.3724e+00 -1.4360e+01 -1.2295e+00 -1.3738e+01 -6.4248e+00
#>  -6.7693e+00  4.9031e+00  7.7290e+00  4.9593e+00  1.1988e+01 -3.7703e+00
#>   1.8442e+00  7.9523e+00 -6.9565e+00  6.0922e+00  3.5425e+00  1.3158e+01
#>  -1.8622e-02 -4.3381e+00  4.6026e+00  2.4567e+00  6.7460e+00  1.2432e+01
#>   1.5490e+00  5.6879e+00  6.3782e+00  4.0045e+00  1.3381e+01 -5.9989e-01
#>   5.1195e+00 -1.3931e+00 -4.8747e+00 -1.0955e+00 -1.0181e+01  1.9294e+01
#>  -1.0241e+00 -8.5806e+00  4.0821e+00  9.5609e-01 -1.0319e+01 -2.3375e+00
#>  -1.1722e+01  6.5177e-01 -2.7847e+00 -5.3115e+00  8.3268e+00  9.3874e+00
#>  -8.5398e-01  4.8621e+00 -9.5259e+00 -7.8092e+00  2.1147e+01 -2.1449e+01
#>   3.0006e-01 -3.7504e+00 -5.4196e+00  1.2243e+01  7.0071e+00 -1.3202e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]