Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  3.4093e+00 -1.8861e-01 -2.0266e+01  1.4997e+01  7.8732e+00 -8.1667e+00
#>  -2.6594e+00 -3.1735e+00 -2.4227e+00  1.8721e+01  6.1879e+00  3.3462e+00
#>  -3.2583e+00 -3.1741e-01  6.8014e+00  1.4984e+00  8.4365e+00 -5.8894e+00
#>   3.2659e+00  1.4548e+00 -1.9906e+01  9.1275e+00 -4.3065e-01 -1.6812e+01
#>   7.8378e-01  1.3715e+00  4.4187e+00  2.4584e+00 -8.0312e+00  1.5392e+01
#>  -1.4074e+00 -3.1942e+00  2.0941e+00 -4.6444e+00 -4.6316e-01 -5.9281e+00
#>  -2.7986e+00  1.3968e+01 -2.6723e+00 -6.0481e+00  1.3153e+01 -2.8273e+00
#>  -2.2527e+00 -4.8134e-01 -2.7743e+00  2.6057e+01 -8.1117e+00  1.0224e+01
#>  -3.6496e+00  1.6641e+00  9.1949e+00  9.7925e+00 -2.0240e+00 -1.5649e+01
#>   3.3770e+00 -4.9918e+00 -6.0450e+00 -5.9153e+00  6.3115e-01 -5.6467e+00
#>   2.3651e+00 -4.0801e+00 -7.7700e+00  9.1688e+00 -1.7452e+01  6.7278e+00
#>  -2.5413e-01 -6.4060e+00  6.1320e+00 -2.1641e+01  1.5343e+01 -5.0190e+00
#>  -6.0961e+00 -2.2665e+00  3.6166e+00  9.2326e+00  4.2633e+00  1.6744e+00
#>   5.5269e-01  3.1717e+00  1.5216e+01 -9.4213e+00  4.4671e+00  1.1008e+01
#>   1.3524e+00  8.5529e+00 -4.2453e-01  2.6677e+00 -5.7519e-01 -1.2085e+01
#>   1.8463e+00  2.7226e-01  1.8399e+00 -9.7480e-01 -9.8918e-01  8.3006e+00
#>   3.7363e+00  3.3194e-01 -1.6125e+01 -3.6385e+00  2.1892e+00 -7.0551e+00
#>  -2.1563e+00 -4.4758e+00 -7.5480e+00 -4.7238e+00 -9.9000e+00  1.0312e+01
#>  -2.3528e+00 -8.2550e-01  2.2492e+00 -3.2341e+00  6.6989e-01  1.1585e+01
#>   2.8296e+00  2.2717e+00  6.7805e-01 -7.2541e+00 -1.0655e+01  1.8026e+01
#>   3.1889e+00  9.3866e+00 -1.0804e+01 -1.2896e+00 -4.8481e+00 -4.7796e+00
#>   2.9266e+00  3.9119e+00 -7.1597e-01 -4.2635e+00 -1.2064e+01  2.3797e+01
#>   1.7239e+00 -1.3540e+00 -5.7897e+00 -5.8072e+00 -1.0105e-01  3.2300e+00
#>   1.8025e+00 -1.0217e+01  5.7439e+00 -1.0927e+00  1.5418e+01 -4.6789e+00
#>   4.5458e-01  6.4249e+00  2.0673e+00  6.8641e+00 -1.4969e+01 -7.7253e+00
#>   7.9414e-01  3.0360e+00 -1.0600e+00 -6.8872e-01 -5.3471e+00 -4.5615e+00
#>   2.0152e+00 -8.7857e-01 -2.9547e+00  1.0414e+01  5.7325e-01  2.0181e+01
#>  -3.5050e-01  7.2457e-01 -9.7976e+00  3.9552e+00  3.0164e+00 -2.3557e+00
#>  -9.2554e-01 -2.4769e+00 -2.3555e+00 -9.9022e+00 -1.0631e+01  7.0639e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]