Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -6.3840e-01  7.0071e+00  1.3971e+01  4.8968e+00 -4.8336e+00 -1.3234e+01
#>  -4.7703e+00  2.7386e+00  8.3850e+00  4.3649e+00  3.8084e+00 -5.8207e+00
#>   3.1130e+00  2.4371e+00 -3.4171e+00 -1.0192e+01 -1.5619e+01  3.4680e-02
#>  -5.9294e+00  1.1263e+00  2.2910e+00  6.7760e-02  4.1584e+00 -1.1588e+01
#>   4.9435e+00 -1.2651e+00  1.3029e+01  1.2075e+01  8.4724e+00 -2.5737e+01
#>   6.1205e+00 -1.7692e+00 -4.1517e+00  4.5397e+00  2.4170e+00  1.4458e+01
#>  -2.4634e+00 -2.0806e+00 -6.3758e+00  9.9950e+00 -5.5635e+00  1.6098e+01
#>  -1.6435e+00 -5.7864e+00 -1.0202e+00 -4.4145e+00  9.0206e-01  2.6295e+00
#>   5.6223e+00 -5.1864e+00  2.7643e+00  6.7065e+00  1.0193e+01 -6.4314e-01
#>   8.7564e-01 -1.4553e+01 -6.4312e+00  4.8478e+00  6.2087e+00 -4.6630e+00
#>   3.8073e+00  4.8286e+00  4.1052e-01 -4.8986e+00  5.9360e+00 -2.0338e-01
#>  -8.2524e+00 -5.8895e+00 -3.2813e+00 -1.0510e+01 -6.2722e+00 -1.7286e+01
#>   2.6952e-01  8.0192e+00 -2.3012e+00  1.1725e-01  8.9857e+00 -1.4570e+00
#>  -3.8824e+00  7.4551e+00  1.3052e+01 -1.1004e+01 -1.3110e+01 -1.2597e+01
#>  -4.7689e+00 -3.7977e+00  3.7497e-01 -8.6646e-01  1.7461e+01  6.4307e+00
#>   1.9464e+00  1.4767e+00 -1.1762e+01  2.4631e+00  1.3712e+01  6.7171e-01
#>  -2.4094e-01  1.0597e+01  3.9641e+00 -6.7823e+00  2.5514e+00  2.0051e+00
#>  -1.9110e+00  2.9255e+00 -6.4973e+00  7.9148e+00 -9.2084e+00  4.0417e+00
#>   2.7779e+00 -3.3004e+00  2.7495e-01  1.0821e+01  1.0438e+01  6.5450e-01
#>   4.1719e+00  5.1090e+00  4.4424e+00 -1.6793e+00  1.2352e+01  5.7437e+00
#>   5.3227e+00  7.0599e+00 -2.8923e+00 -2.2207e+01 -5.2314e+00  2.1065e+01
#>   3.0503e+00  8.6644e-01 -6.5384e+00 -6.7636e+00 -2.3684e+00  3.5472e+00
#>   3.4848e+00  3.4683e+00  5.4743e-01 -1.7362e+00  1.3411e+01  5.4663e+00
#>  -5.3593e+00  1.9327e+00 -6.3721e+00 -4.4773e-01 -2.6109e+01  1.2974e+01
#>  -3.1484e-02  4.9579e-01  3.9247e+00  1.1896e+00 -7.7224e+00 -8.4465e+00
#>   1.4926e-01  7.5090e+00 -9.0295e+00 -1.2009e+00 -8.8624e+00  2.2618e-02
#>  -2.6064e+00  4.9816e+00  1.3691e+01 -2.9766e+00 -3.6852e+00 -1.0116e+01
#>  -8.9572e-02 -4.4183e+00 -1.6188e+01 -1.1292e+01  2.5385e+01  6.0564e-01
#>   1.8047e+00 -3.2690e+00  4.5937e-02  2.0798e+01 -4.0223e+00 -1.4672e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]