Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.4361  -3.8399  -7.7339   9.4252   3.7653  -7.2446   5.3059   1.9831
#>    5.7342   4.6398   4.0207  11.5283   2.3162  -2.5326   7.0058  14.8134
#>   -1.5808   0.9571  -1.0319   9.3286 -17.3584   4.4741  -0.4857   5.4812
#>    0.2720 -13.6094   3.7023  11.1266   8.1676  -0.7394  11.1598  13.2093
#>    1.6447   1.0514   2.1751  -2.0097  -1.7650  -3.7594   6.7439 -18.6293
#>   -1.6792   2.2285   1.2541   1.2369  -8.9284   4.1494 -15.5969 -24.1898
#>    5.7314   5.6835  -5.4462  -5.6868   3.4510  -9.4036  -3.1415   5.8659
#>   -2.5532   4.8683   8.5311  10.6955   2.3085   1.6984  -4.1996  -7.0875
#>   -3.4796  -1.9527  -9.4036  -3.5669  -4.4361  10.8740  -9.0197  -1.2544
#>   -1.9123   1.3368   4.6510  -0.5264 -16.2336 -15.6924  -4.6552   1.5661
#>   -0.1984  -4.0281  -3.9435   1.6007  -6.6739  -3.0108  10.3557  13.5860
#>    4.1897  -7.5794  -0.1182   4.9204   5.4916  23.2288   3.5696  -3.6617
#>   -3.7335  -0.9122   5.1383   2.7638 -10.7170   6.5977 -10.1465  -8.6224
#>    0.8672   6.4617  -0.9985  -5.3704  12.3801  12.0827   3.1634  -1.4112
#>    3.9799  -3.0733   3.8222  10.8943  -4.3742  -4.4303   4.1388 -19.0648
#>    9.5008  -5.9180  -3.5738   0.4061   5.0451  -7.0623  -6.8957  14.2619
#>    4.6076   3.6659  -9.5999  -7.5785   2.5593  -4.3666 -13.5186  -5.4349
#>   10.9847  16.7364   9.1229  -3.8927  -4.1624   1.5073 -13.4044  -8.3235
#>    4.7335  11.6239   9.5878  -1.6625  -7.7758   1.8690 -10.3431   8.4720
#>    2.9434  -0.5188  -3.0495  11.6733  -0.8921  -1.1362   6.3497  11.9762
#>   -5.5354  -6.8038   2.2715  -8.8899   5.7423   2.3878  21.7541 -15.6668
#>    0.0253  -3.4959   0.0788 -12.7872   2.4631  -4.8369  -5.2055   9.5272
#>   -0.4176   0.9259   2.4601   3.6287  -3.5290 -15.6409 -13.8813   3.5843
#>   -8.2404   1.6537  -3.8592   0.3284  -2.7286  -3.7666   2.2126  -4.1562
#>    3.4156   4.2793   0.0593   8.5801   3.7346  -9.0056  -2.0235  11.1711
#>   -6.1142   3.8410   6.7599  11.9960 -21.0840 -15.6662   0.5291   4.5843
#>   -1.8582  -2.6476  -2.5186   8.9495  -2.0701  -6.7328  -9.1137  -6.9569
#>   -0.8468   1.3654  -3.1529  -7.8803 -11.2570  -4.1123  17.0684  13.4925
#>   -9.9071   1.0707  -0.8664 -11.6669  -9.0368  12.4006 -19.7241  -4.2634
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]