Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.4069   1.5958 -11.9902  -4.3551  -4.4161  -2.6565  -7.8970   1.3455
#>    1.9827   7.4095   4.4256   5.1252  18.8808 -14.3913   2.5270  -2.8889
#>    3.1824   4.7686   1.3432  -3.6317   1.9980   6.8362   3.1604  -7.2109
#>   -5.1698   2.8841   0.2459  -2.3176  -0.2145  -7.5285   5.8877   5.3691
#>    5.9197   6.0571 -15.4985   4.4383  -6.2829  -0.5072  -6.3388  -0.1594
#>   -4.3490  -0.7253  15.6309 -19.4849  13.7199  -2.7654   0.5138  13.3337
#>   -1.0639  -2.0685   0.1092  -1.0623   1.8722 -12.6029   2.6050   9.1250
#>   -1.5705  14.5362  -7.5795  -6.0646   9.3581  -8.7132  16.7124   4.1588
#>   -2.3817   0.6339   1.3781  -3.4253   0.6249   2.3786   2.5322   1.2257
#>   -6.3524  -1.6594  13.8579  -6.7658  -0.6642  -4.9769  -3.7824  -0.3973
#>   10.1579  -5.2759   1.1900   4.1583  -4.7674   7.0471   5.1104  -3.3725
#>    1.0971   3.4940  -6.8475  10.1486 -16.0044   8.8103 -16.8294  15.9013
#>   -7.2714   2.1255  -7.8890 -14.9423  16.8020 -12.1682   3.1516  -2.8074
#>   -0.5058   6.4743  12.3054   4.3864  10.9591 -12.0349   5.5259 -16.1613
#>    2.3333   0.0493  -8.4990 -12.3152  18.6529  21.3699  -5.6881   0.5604
#>    1.1048   8.1104  -1.9443  10.6447  21.2540  -3.3898   8.8838  -1.1478
#>    9.3888  -1.4110  13.5106   6.9534  -7.4816   3.8708  10.6103   8.4231
#>   -2.3449 -14.8177  -6.5031   2.3741 -21.2465 -12.7945  15.1774  -4.7502
#>    1.9470  12.0755  -1.2756   6.2364 -17.7968  -3.1961  -8.3638 -11.6748
#>   -2.1119  -1.8616  -1.0469  -4.9140   8.9850  -9.9748   4.6549  -5.1041
#>   -2.5907   6.8821  -5.3208  -7.2755   0.7633  -6.9141  17.7064   5.7877
#>   -1.7879   0.4561   1.8152   9.4376   5.8358  -7.0087  -4.6344  -2.3978
#>    4.6658  -6.3705   6.0795   5.9975   2.3470   3.9157  -9.2491  -4.7105
#>    1.3118  -3.1959  19.3641  -7.5510  -8.9032  -2.6288  -4.3433  -7.5361
#>   -4.8582  -7.2087   6.4556 -11.8446  -1.3546   1.7313   4.6447 -19.8586
#>   -2.1341   4.8301  -4.5261   8.2377  -4.0235   4.9116  -1.5513   8.6665
#>    2.3740  -4.0069  -6.2334 -10.3339   5.2283  10.0080  -6.6162  -5.5805
#>    7.1963   9.6635  -7.7005   0.8361  -5.4799   8.5087 -11.3508  -3.4043
#>   -1.0544   4.0671   3.0405  19.0949   1.4625  23.3967 -12.6771   2.7207
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]