Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   9.6653   3.3656 -11.8227  23.5459   5.9670  -7.1278  -4.8407  -3.7647
#>    0.5195   2.0344 -14.2052  -5.4562   6.6235 -10.3181  -2.9425 -10.4911
#>   -0.6772   9.6553  -4.9221   9.2988   0.8787  -5.8810  -2.6469  -5.2092
#>   -1.3125  -1.0332  -0.0560   1.0819  13.2345  -4.4114  -5.0683  11.3501
#>   -1.0942  -1.3568   0.2471 -11.8541   3.7184  -0.7357   5.9266   6.7910
#>   -2.5168  -7.2199  13.3034  -4.0227  -0.0700  14.1830   8.7090   4.6426
#>    6.2957   9.9589  10.5578   6.6982   3.4392  -1.7862  20.0822 -12.9127
#>    0.4119   7.4922  -5.1610  11.5130  -3.8086  -9.2417  11.2654   0.9010
#>    1.8783  -2.7920   7.5345  -0.0159  -1.8425  -0.1847  -2.0809  10.4796
#>    0.1390   9.3721  -6.8465   3.8717  -4.3500  10.6521  -0.6249  -2.8441
#>    1.5660  -4.6227   5.5437  -9.0098   3.7011  -1.2738  -0.8331   2.3744
#>   -2.4659  -4.1571  -4.0466 -10.4845 -21.0327   6.2618 -11.5656  -4.4261
#>    2.5364  -0.3609   4.7750   5.4153   3.6346  -7.2119   4.4354 -17.1914
#>   -6.3935   3.6239  11.2739 -11.5032   6.5349   4.2297  -2.3643  -0.6423
#>    2.7837  -4.9116   0.2021  15.7267   2.6509  -0.2580   7.7711   6.5929
#>    4.0166   0.8231  -1.6841   7.0509  -5.2489 -16.9011  -2.5607  -5.3008
#>   -4.0733  -6.5117  -0.7322 -18.4644   0.7440 -10.7839  -3.3910   8.8695
#>   -1.4594   7.1792 -10.0475  -6.9080  -2.2563  -5.5278  -3.7230   6.4305
#>   -0.4834   0.1049   1.6032   4.9549   4.6643 -19.0961   9.6580  -9.7980
#>   -2.6688   1.6471  -9.7061   7.2777 -10.3593 -10.7903  -0.9725  -1.7964
#>   -3.6166   0.9758  -0.8767  -6.0083   7.1390  -3.0076   6.1526  10.2707
#>   -2.3155  -1.1310   1.2891  -5.1844   2.0081  -8.8735  14.6079   2.1393
#>    0.1534   4.1468   0.8301   8.2230  -9.6515   5.6568   0.3120  -3.3100
#>    3.3196   0.4998  -2.9130  -6.1559   8.6514   4.2222 -21.5330   0.8645
#>   -2.9825   1.6959   8.0331 -12.3720  -8.3188   1.8677  -8.8235   6.0549
#>    6.9676  10.7318   1.6162  -0.1472   5.0077  -3.5720  -2.5097 -17.0507
#>    5.5722   1.9013  -9.7941  15.4952   4.5216  -1.5536  -6.5776  -6.3974
#>   -2.0473   4.2317  -7.7017  11.1616   8.2201   9.9748 -12.1153   5.7255
#>   -1.6571  -7.3886  14.3966 -12.8741  11.9154  -4.9704  12.8680  31.3558
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]