Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -5.2347e+00 -2.5196e+00  4.1546e-01 -9.9587e+00  5.7075e+00 -7.1973e+00
#>  -2.8183e-01  2.0181e+01 -2.6754e+00  4.3406e-01  4.3185e+00  7.6977e-01
#>  -1.0663e+01 -1.5582e+00  6.7289e+00  1.1244e+01  7.2922e+00  6.7708e+00
#>  -3.0221e+00 -7.1701e+00 -6.5885e+00  8.6716e-01  1.6060e+01  6.0869e+00
#>   1.3777e+00  2.5599e+00  1.2408e+00 -6.0083e+00  2.2250e+01 -1.0527e+01
#>  -5.0592e+00  3.5439e+00  3.2380e-01  1.3937e+01  1.1968e+01  7.7863e+00
#>   3.1828e+00 -2.4008e+00 -7.6415e+00 -2.0344e+00 -1.4704e+01  3.6068e+00
#>   1.5243e+00 -2.5096e+00 -5.7792e-01 -5.3594e+00  5.1045e+00 -1.1178e+01
#>   9.4416e-01  6.4945e+00  9.8375e+00 -7.1076e+00 -2.0590e+00  4.4867e+00
#>   2.1727e-01  1.1407e+01 -6.5479e+00 -2.6464e+00 -7.4669e+00  2.6864e+00
#>  -1.1490e+01 -8.1979e+00 -1.5557e+01 -1.2075e+01  1.2611e+01 -4.7014e+00
#>   3.2753e+00  1.8140e-01  6.3787e+00 -6.3972e-01 -5.0340e+00  4.1345e+00
#>  -3.2774e+00 -4.7536e+00  5.5767e+00 -7.0639e+00 -6.0848e+00 -1.1870e+00
#>  -3.4385e+00 -5.9057e+00  3.1751e-01 -4.6755e+00 -9.8274e+00 -1.4150e+01
#>   2.8938e+00 -4.3146e+00 -1.0283e+01  3.7090e+00 -1.7970e+01  1.3192e+01
#>   4.2197e+00 -4.6668e+00  7.8507e-01  2.1123e+00 -5.3107e+00 -9.3013e-01
#>   9.0454e+00 -3.7541e+00 -4.4241e+00  1.8164e+01  7.5164e-01  3.9617e-01
#>  -4.7787e+00 -1.5834e+00 -9.6204e+00  3.4161e-01  6.1175e+00  2.2599e+01
#>  -5.9615e-01  1.1790e+00 -2.6512e+00  7.0037e+00 -2.4654e-01 -2.8364e+00
#>  -3.7492e+00 -3.7209e+00  3.5973e-01 -2.7949e+00 -5.4167e+00  2.0760e+00
#>   4.1155e+00  3.9173e+00  7.2410e+00 -2.1225e+01 -7.1744e+00 -3.2224e+00
#>   6.3476e-01 -6.8807e+00 -1.4945e+00 -9.6118e+00 -1.2611e+01  6.4533e+00
#>  -3.6184e+00  1.2567e+00 -6.6504e-01 -5.1640e-01  9.7686e+00 -1.2905e+01
#>  -6.6416e-01 -7.6930e+00 -7.1994e+00 -8.6872e+00  9.8685e+00 -1.3263e+01
#>   3.0843e+00 -4.4916e+00 -3.3951e+00  4.3646e+00 -2.9067e+00 -5.8415e+00
#>   1.2070e+00 -3.8855e+00 -5.0872e+00 -4.5276e+00 -2.2958e+00  5.6753e-01
#>  -3.9316e-01 -9.7158e+00 -9.7414e+00 -5.8988e+00 -1.9036e+00 -4.5719e+00
#>   1.8086e+00  2.8553e-01  6.1826e+00 -2.4986e+00 -3.6696e+00 -3.7141e+00
#>   5.6264e+00  1.0897e+01 -2.7250e+00  1.7446e+01  4.1608e+00  3.9618e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]