Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.1846   2.1082  -4.3608  -0.6103  -3.9968  -2.7289   0.2392  17.3959
#>   -1.1519   9.0125   1.1660  23.8058   1.5315   5.2029  13.2267  -5.3176
#>   -1.8957   2.5959   1.9513  -1.5743   2.0898  -3.3405  -7.2051   2.5931
#>    2.6067  -4.4598   6.3200 -17.9216  14.3573 -12.9166  -8.9099   7.9642
#>   -3.3036  -4.7368   4.6093   1.2027   5.5003 -11.0032   6.9644  -9.3420
#>    4.5050  -9.9446  10.8225 -14.0134   4.8468  -8.6393   3.1214  16.5243
#>   -3.8543   2.6202  13.0309  -5.0899   5.7598  -3.0349  19.2717  -7.9569
#>   -1.8980   0.7456  15.8677  -3.0084   1.7248 -11.2637   5.6460   1.1733
#>    0.4606  -5.6086  -4.0153  -3.0136  -3.0148  -0.8351  13.3341  -1.8285
#>   -4.1905   0.9582 -10.7116   2.7192  -0.1828  -3.7244   6.4629 -12.1007
#>   -1.1681  -7.5866   5.0155  -2.8647 -16.2766 -10.4881 -22.8296  10.4931
#>   -2.3625   2.0954  -9.8945   2.1916  25.8409  -0.3850  -1.3307  -1.6027
#>   -4.2780  -5.9221   1.9044   2.1246   6.8400  -5.1529 -12.5224  13.2969
#>   -4.1160  -1.4245  -8.6233  -1.1619  -3.2484  -2.7945 -12.5308   0.1090
#>   -1.4535  -1.1839   2.9930 -12.1395   3.6652  -4.1893   9.9419 -12.1663
#>   -1.8704  -0.4654   1.9122   6.2370   8.9198  21.7969   8.9311 -15.9926
#>   -0.6374  -0.9791  -5.6281   0.6921  -6.2554  11.6621 -11.0218   5.4438
#>    2.2787   8.6274   7.7893   5.3933   2.7829  -8.0098  10.5347  -5.6514
#>   -3.7293   0.8585  14.9344  -3.2929  -6.4396  -6.0798  -8.8407  -2.4284
#>    0.1596   2.7399   1.3089   5.4233   3.9649  12.6743  -9.5912   3.0761
#>    1.7176  -3.5118   4.2285 -10.6875   1.9906  -8.2684  20.0928  18.0705
#>   -0.6582   0.5719  -4.6111  -5.9819   0.6827 -23.5049  18.4146 -11.5200
#>   -3.9582  -4.1585   2.4569 -14.6567  -9.6586  -5.7900   1.1937   1.2915
#>   -3.9562   0.7574   3.1818 -15.1085  -1.7406  -1.7594  -0.2379   2.1736
#>   -5.5280   5.8623  -7.5538 -15.6037  -8.1945  -1.7377   8.6971  -2.1437
#>    0.6603   3.0666   8.9127  -8.1324  -7.0227  -0.6258  -2.0384  19.6753
#>   -2.7527  -8.7100  -9.8322 -10.8773   0.5723   5.7720 -11.1125  -0.4925
#>    1.5165  -5.2330  -3.8422   2.2205   8.5983  12.3759  -4.3668  -0.6712
#>   -1.2515   3.5614 -10.9233   0.0145   2.3296   5.8816 -10.8904   3.6382
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]