Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -7.5008   4.7224  11.2369  -5.8807   3.2188  11.6994  -5.9791   0.6442
#>    1.3208  -9.4188  -1.4878  23.5985   4.5127  -2.9738   1.6576   8.1225
#>   -2.1746  -6.3641   5.9015   4.4682  -7.2784   0.1544  12.4856  -5.3545
#>    4.0466  -8.3250  11.8056   1.1594   6.2600  -9.2637  -0.5798   3.1347
#>   -0.0462  -5.4555   3.7328  -1.3038   1.8812  -8.7518  -3.8664  15.2594
#>   -1.6379  -3.1121  -7.3745  -7.2224  -6.8614   6.8208  -3.0817   6.6676
#>   -1.3808   1.4106   6.2025  -3.8607   4.1607  11.3635  -8.8075  10.9925
#>    0.3235  -5.1554  -2.4394  -3.0612  -7.4701  14.6832   9.9061  -5.0398
#>   -4.5531  -5.6749   8.6299  -3.7183  -3.0710   2.7148  15.2602   6.4057
#>    0.1471   2.0608   4.2808   4.8618 -24.2929  -5.0716   3.4237   2.6026
#>    2.2077  13.0484  -3.5717   4.8627  -0.5302 -13.0344 -15.6116   1.0873
#>   -1.0763   1.1599 -13.9525   4.8033   9.0477 -11.1718  13.0170   1.5248
#>   -3.9314   6.0214  -0.6373   4.2105   4.9159  -6.7412  -1.7703 -14.5106
#>   -1.6620  -1.1942   4.6642   0.1556   6.7184 -11.2314   0.3905  -1.6039
#>    3.8861   0.4608   7.0333  -2.3908  10.1368  -5.4864  -5.3085  -3.3753
#>    6.6308   9.4170   0.1887  -1.7376  14.4192  -6.5407 -11.9928   8.9992
#>    3.0592   6.2801  11.2766  -2.6507  -4.6665  -7.4112   1.2756   7.8878
#>   -5.3894  -0.3236  -6.6814  -4.8290  -9.7288  18.4674   3.7551  -4.9424
#>   -1.7782 -10.8958   3.7884  -8.5142   0.4716   2.8595   5.2994  -0.8490
#>    2.3609   4.3438   3.0982  -0.6478 -13.1296  23.2936 -11.7957 -11.6942
#>   -4.9814  -5.5792  15.4883  10.5223  12.2031   5.8292  -3.5854  -1.4287
#>    2.8300   8.3656   1.4426  -0.3959  -7.5251 -13.0804   1.4462  -0.7770
#>    1.0517  -2.1613  -0.9264  21.2301  13.0663  -5.8754  -5.9436   5.5084
#>    3.9265   5.2787   2.7256 -20.9194   7.4307  16.5944   2.1248   4.2147
#>    2.7466  -0.8821  -2.1997  -0.5472   8.5645  -8.9114  -6.8147  -7.3274
#>   -1.0976   5.0537   9.8404  -1.6495  -2.3992  -5.9879  -7.8839   2.4456
#>    4.5540  -3.9511   0.0114 -18.0243  -2.7091  -2.8250   6.6145  18.4886
#>    2.4440  -4.3508  -4.9881  -3.9736 -13.8338   0.2289   6.4163  -7.7529
#>   -5.0450   0.4648  -3.4381 -10.8546 -12.5741   8.2646 -13.5146  -3.8756
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]