Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.7615 -10.3724   2.5918  -4.2458  20.4842   2.5449  -0.1214   7.6639
#>    4.1959   6.9719  -2.1661  -1.2325  -2.3072  16.0026 -16.0097  -6.1950
#>   -0.9731  -3.0599  -8.0936   1.1693   3.3781  -0.9027  -7.0219  -6.5259
#>   -3.3619 -12.2467   1.7789   0.1934  -3.1233   5.1354  14.1384  -6.1926
#>   -3.9254 -12.0433 -11.3465   4.9737  -2.7489  -2.9139 -12.3764  -4.6231
#>   -4.0670  -8.1698  -4.2024   9.5681  -1.3956   9.1083   1.6202  12.0183
#>   -4.0563   3.4481   0.0032   2.2186  10.6359 -11.1234  -6.0581  26.1743
#>   -3.3135   9.1668  -4.8150   1.0088   1.4345  -0.9037  11.4661  12.2311
#>   -1.5995  -3.1275  -0.8040   4.4033   8.6041  11.6988 -30.9005   1.5181
#>   -2.7154   3.3909  10.0449  -4.4147  -3.4544  -3.7545  -3.1515  16.5501
#>    3.1665   1.1328  -1.6493   1.7739 -20.9296   0.4218 -14.0268  -8.9331
#>   -4.6974  -0.1299  -0.8784  -1.5796   0.4336   4.1968   5.7442  17.8213
#>    0.7973  -0.8452   6.1144  -9.0590   5.2691   4.1956   9.8539 -13.7681
#>   -4.4649  -0.6973  -2.0215   8.6723  -8.8235  11.2248  -4.4054   0.6724
#>   -2.0067  -1.5615  -2.9533  -0.7823  19.5451   4.3373 -15.9131  -0.0841
#>   -0.0459  -0.3720  -7.5212   1.5717  -6.0855   0.0045  -9.4465   8.8423
#>    0.7797  -4.6037  -9.6382   7.2010   1.1722 -13.1047  -2.6541  -7.1100
#>   -2.6459   4.1548  -3.3174   3.0130 -10.7229  -3.7390  17.6950   9.2038
#>    4.2519   1.8686   2.8161  -9.1897  -2.6429 -23.7466  -8.2534  -7.9426
#>    4.7803   7.0570  -2.6928  -8.4055 -13.4280  10.5560   1.2942 -15.6061
#>    2.5813  -2.7171  -4.1950  10.9949  -2.2720  -7.8160   5.0474   9.3042
#>   -6.1004   1.6789   4.7274   6.5094   9.4582  -0.7679  -5.0186   9.5083
#>   -0.4859   7.3468   4.2465 -29.3634 -10.7618 -13.0166   9.4089  -4.3065
#>    0.4244  -0.3799   4.9399   2.1012  20.4925   5.2810 -15.1440 -17.4740
#>   -0.7773  -1.8570  -5.4736   2.8313  11.0157   3.6294   1.2897   6.0293
#>    3.3803   4.5400   3.5156  -1.3073   0.3848  -1.7046   8.1786  -9.3417
#>   -2.9000  -7.3911 -11.9520   4.5212   0.1198 -18.1392   6.8979   5.9306
#>    2.8358   3.0981   0.1475   3.0711   6.7738   3.0780 -19.8308   4.1748
#>   -2.0991  -5.0910 -13.2178  -9.6133   8.0104  15.9129   2.9324   9.7785
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]