Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.7535   4.7114   2.7074   6.0318   4.5974  -6.6074   7.9908 -27.1181
#>   -1.8339   7.4414  -1.6578  10.5829  -1.6296   5.8104  -5.6152  21.4527
#>   -0.0322   2.7602   0.8079   1.6206  -0.1707  -0.2502   6.2772  -1.5904
#>    1.9711   5.0590  10.5583  18.8306  -0.1240   0.2025   1.1675   6.5374
#>    3.2129  -8.3017   5.6933  -7.1637  -5.1845   9.9725  11.5410  16.5290
#>   -0.9770  -5.8769   0.8037   1.3916   3.9799 -12.5100   6.2582 -15.5874
#>   -1.7212 -10.7188   2.4199  -5.1408   3.3206   6.1173   8.8196   2.2981
#>   -1.7149   0.6189  -8.5917   1.2069 -11.8687   7.4407  -1.0759   7.1240
#>    8.2240   0.3491   9.7072   1.3570   7.5148   4.1230  -4.2041  -8.3388
#>    0.9594  -1.6432   6.0545  -9.4457  10.6236   1.6246   6.0862  11.7373
#>   -1.4542   1.6817  -0.5029   3.6653   7.5899   3.5458   5.4277  -6.6897
#>   -0.0876  -5.5544   9.9691   0.0179  -8.5772   5.1113 -10.1989  -4.9769
#>   -1.6978   1.9987  -3.4392   4.6079 -13.5705   1.6492  -0.3855   1.6150
#>    0.0981   0.1251  -4.8948  -1.9431   3.5829  -9.4775   8.5626  -2.0784
#>    1.9734  -1.3689  -3.0532   9.8002 -19.2359   7.7844  11.7390  -0.7013
#>   -4.3776  -3.6154   1.3087   2.0215   6.8054  -7.9861 -12.2687  -3.6791
#>   -4.5820  11.3086   1.8400  -9.8868  -5.8092 -10.3161 -14.3258  11.1496
#>    2.1398   0.9337   0.1784   6.8917   5.3504  -2.9129   2.5658  -2.6115
#>    0.1443   2.6632  -9.3956  -2.8324   7.7924 -12.5473 -20.4339   3.4929
#>    2.2956   7.9871  -2.5416   0.3716  -0.9224   2.4493   4.2996  -2.5862
#>   -0.5250  -3.2623  -5.2517 -14.2761  -1.6315  -8.6632   0.6454  -6.3234
#>    3.2942  -8.5166   6.6367   8.1794  -0.0544 -18.9291  -3.7229 -18.3070
#>    3.4677   3.6734   4.8319  -3.3021   5.8704   9.9556   8.2217  11.5321
#>   -0.0284   3.0485  -2.7551  -3.4888  -1.0469  -1.5334   3.0242  12.5685
#>    1.1479   0.1458   3.9890  -5.4858  -0.4388 -22.2487   7.1699  -1.3409
#>   -4.5891  -0.8184  -2.3471   9.6911   1.2051   5.1789   6.2856   4.9919
#>   -0.6866   5.0529  -3.5141  -2.6725  12.7199   2.5651  -1.2151  18.8591
#>   -0.6582  -8.6082  -5.4793  -4.5055   1.3821  -7.6318  12.7924 -14.2343
#>   -2.9156   3.7698  -3.9202   5.7145  10.6539   4.3567   2.2307  -0.6160
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]