Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -6.1186   5.4199  -3.0534   3.4709   4.4704  -0.0563   3.1842 -14.9494
#>    0.2022  10.5129  10.1725 -19.0167  -0.7766  -0.1618   3.2674   0.6587
#>   -7.3157   3.6867  -6.0104  -0.8271   1.4799  10.9929  -5.1432   0.3007
#>   -0.3072   2.5851  -5.6101  15.0191 -11.2143  -0.4406  -2.0321   9.4026
#>    4.8533  -2.3209  -2.4071  10.3007  -4.1743   3.2818  -8.6364  -3.5762
#>   -0.3950 -10.1989   6.8453  14.6717 -11.3081  -1.4535   9.2251  -4.5125
#>    0.2625  -3.4612  -0.0869  -3.5923 -15.0509   3.7201   1.6006  14.3593
#>    4.6045  -8.2995  -4.8493 -16.3986   3.1602  13.1108  -2.7390   5.6508
#>    0.9438  -1.9903  -2.3422  -1.6555  11.9162  -3.8159  -6.7310  -2.8288
#>    2.7246  -0.7875  12.7595  12.4491  -8.6992  12.8068  -6.5940   5.2083
#>   -0.8645   5.6987   7.4126  -1.9255  15.0340  -5.4733   5.2583 -16.4016
#>   -6.4253  10.7743  -1.2629  -8.8505  -8.1810   6.9763  12.0797  -9.3601
#>   -1.4958  -6.5817   5.3906  19.6528   0.1691 -11.6843  -1.4510 -10.0539
#>   -1.2256   0.5434  -1.8465   7.8853  -5.2683   5.2066  -0.3678   4.4074
#>   -1.6743   4.9056  14.1859 -12.2544  -1.6570  17.5183   7.9244  -5.8266
#>   11.2814   3.6792 -18.4022   4.4325   6.9037 -11.2332  21.8238  -4.0276
#>   -1.1603  16.8879 -12.5158  -9.2206  -8.4960  -3.1367 -18.0395   0.4510
#>    4.0526   0.3205  -9.5540  16.2417 -12.9932   0.6303  -0.9679   3.2285
#>   -2.4666  10.3371  -7.0968  -9.1011  11.6445  -4.0422  26.4431  10.4774
#>   -3.4687 -14.8336   0.1018  16.5165 -15.5855   1.0385  19.0067 -10.9891
#>   -2.2131  -4.9722  -3.0793  13.5365   6.1415 -30.3025  -0.0012   2.1109
#>    0.0690  -4.2986  -9.6285 -15.0881  -0.7548   6.8133 -12.7991   5.7304
#>   -0.1213   3.7882  -5.2363  -4.1649   8.3437  -6.3626  -2.5683   4.6203
#>   -3.9506  -2.4828   4.1456   5.4285  -1.3864  -3.8196  -5.3546  -4.0950
#>   -3.7756  -5.5850 -10.3655   2.2506   4.6581  -5.7913   6.4459  -5.9900
#>   -3.2758  -7.3872   2.1549  12.5729  -3.1116   1.2127   5.3801   2.2133
#>    4.4976  -8.1566  -8.7105  -3.5378  -5.8015  11.9232  -7.9570  15.2474
#>    0.0854  11.4124  -9.3254   3.6570 -11.8583   8.7104   1.7073   5.2522
#>   -2.7917  -2.0972 -12.3930  -1.5031  13.7898   0.3863   3.4659  -2.8265
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]