Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.0826  -3.2328  -3.1780   0.0298  -8.2496   5.5252   3.8265 -16.0103
#>   -8.5469   0.1609   8.9542  -5.2216  19.1733   6.7107   6.8135 -15.0976
#>   -2.9840  -7.8190  -8.3076   2.8882 -16.1129   1.2617   0.8731   2.6579
#>   -1.0351   6.5417   0.7302   9.6141 -14.6469  -8.6066   4.2328  -5.5201
#>    4.9492   9.2436  -1.0317  -2.8284   6.6090   5.7863 -24.2305   0.0088
#>   -1.9106   3.9358  -6.7325   6.2673 -10.3832   1.3956   4.0447  15.1818
#>   -2.1144   2.8540  18.4233   8.6695   1.7338   9.5815   2.0380   1.4900
#>    5.2017   0.4362  -4.8418  -1.2964   0.6606  -3.5772 -12.4188   0.3288
#>    3.1737   0.4119   2.6829   2.9172  22.7073 -10.6735   6.8735  -5.9445
#>    2.1190   2.3197  -3.2908 -15.7091  11.0776   7.3606   2.0950  -3.0806
#>   -0.3880 -15.1609  -8.9384  -3.4452   6.9085  15.2111  -3.5011   9.9428
#>   -1.3822  -6.1490  -6.0171  -0.3526  -1.9978  -0.7748   5.2712   9.7836
#>    0.5850  -2.6186   1.2230  -7.9513  -7.2304   4.6257  -3.8087   5.1504
#>    5.0742   3.4272   0.1988   5.0003  10.3331   0.3624 -17.0828   4.9309
#>   -2.1510  -6.3021  -3.3675   0.2285  -4.2040   6.4637  -7.7324   6.1154
#>    0.4576 -10.4537  -9.5633  13.5898  16.5553   8.1340   9.4123   5.8942
#>   -1.3384 -11.5494  -3.2358  -7.8198   1.6578   4.1971   5.2327 -12.0638
#>   -0.7149  -1.4571  20.1630  -5.7633  -4.3603   3.1665 -10.3730  14.2092
#>    8.4559   5.7431 -13.6568 -15.2266   6.5101   3.5963  -7.4127  -5.1881
#>   -3.6863   3.6167  -6.5865   3.9422  11.6411   5.9919  -2.3535   6.7854
#>    2.2983   1.1177  -3.0682 -11.6921   0.8689  11.6850  -5.7185  -8.9297
#>    8.5941  -5.6829 -20.0004   1.0551  -3.7755 -18.6705   0.8061   5.8681
#>   -8.0366   3.6885  -2.1077   7.3701  -8.4471  -3.3633 -10.6362  10.2637
#>    0.3454 -15.4397  -4.5829 -10.9853  -5.3428  12.4051  -5.9169  15.0745
#>    3.9859   4.8788  14.9213 -15.5092 -20.9979  18.4543  -0.2538 -15.4775
#>   -3.2215   4.1086   6.8447   8.8186 -14.7425   9.6034   0.9043   6.4579
#>   -1.8873   6.2358  -0.2238  15.5462  -7.2763 -10.3539   4.6643  -5.6230
#>    0.9980   8.0700   8.9527  -9.4624  -3.3798   5.6563   1.8500  -5.8979
#>   -3.0767  -3.6875  11.8195   2.3209  12.3317 -10.0017   1.0081   6.3692
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]