Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.5980  -3.0563  -7.9584 -17.4208  -0.3055   1.6592   7.9775   1.4991
#>   -3.9616   3.4707   8.8230  -2.6082  -2.2662  -0.3638  -0.3463  10.4326
#>    8.3839  -9.6185   8.0999  -9.7967  10.2439  12.1746  -1.6760   3.6738
#>    3.3537  -5.6087   4.7837 -14.2077 -10.1269  -5.7558  -7.1688  -8.9686
#>    7.9659  -2.2931   2.8683   8.5075  -1.8842   6.8373  -5.9178   5.0950
#>    5.9174   0.7196  -6.6726   0.1820  -0.3065  -3.9426   8.4883  -1.6252
#>    4.6434  -0.9503  -6.8610  -8.1404   6.8009   4.6698  -0.7345   9.6339
#>   -6.8774   5.4036   4.8836  -5.1881  -1.6537   5.5972  -2.6631  -6.0786
#>  -10.8508   3.0724 -18.2150 -15.9395   1.4009 -10.4154  12.6338   6.4895
#>   -2.5495  -2.9657 -14.5345   5.2747 -12.9157  -3.5277   2.0397   7.4693
#>   -2.2158  12.3135   4.9589   1.2542  -4.6445  -6.8126   6.8851  -1.4374
#>   -8.6058   2.8832   4.2372  -1.6129  -3.5685 -14.5869  -3.6318   2.5334
#>   -3.6495   6.9835  -2.3691  13.7289  -2.4591 -22.1734   4.5594 -10.1076
#>    0.6966  -2.4047  -9.2651  -1.4898 -13.7223 -15.8146   4.0242   0.0177
#>    3.7243  -8.2137  -0.9990 -12.9395   1.6662  -6.3816 -11.0215   5.5688
#>    3.4874   2.3306   4.7631  13.4411 -13.3946  -6.0494   4.6987  -5.6212
#>   -5.6156   9.4895  -0.7686   1.3841  -2.2077   2.8916  -0.5691   2.3205
#>    6.3193   6.0106  -2.1819   1.7708  10.1356   0.6304 -10.1655  -0.3481
#>    5.0038  -4.8695   6.6223  -5.6015   8.3954   5.1156  -1.2450 -14.9693
#>    3.5266  -1.4662  -0.7179   3.0835 -16.1202   3.0835   3.2237   3.2463
#>   -2.1614  -5.0320  -3.0014  -4.5394   4.9455   7.4476   5.5685  12.1113
#>   -3.4298   6.6995   1.2868  -4.2120  13.4561   9.2339  -0.1440  -1.8798
#>    2.3165  -7.1576  -5.8494  -9.2518  -1.5874   8.3038  -7.9400  12.4129
#>   -7.5453   9.7540  -0.8044   7.3002   1.9777  -5.1075  11.7922  -2.3923
#>   -5.0421  -3.4117  -6.3164  -7.2947  -8.4781 -15.7749   2.6463  -3.7480
#>   -1.3790  -6.3756   3.4449  -9.2178  -2.7049  -9.2711 -10.8934 -12.2856
#>    0.9691   6.7357  -6.2167  -7.5232   4.8174  14.0012  -7.5636   1.9386
#>   -1.1709  -6.8730  -7.7278   0.4932   2.6129  -3.0776   2.4116   0.2322
#>    1.6823  -3.3158   6.0971   5.1036 -12.9633   0.0409   2.3465  13.2062
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]