Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.2475  -4.3523   0.2644  11.3386   3.5122   5.1553 -19.2092  -2.0945
#>    2.0292   7.3212   7.1231   9.7267  -8.7471 -11.5030   1.1518   6.0373
#>    1.8058  -1.3385   8.3391  18.3395  -3.7198 -11.5519  -7.0451   8.5825
#>   -0.7661  -5.6103   9.3046 -10.9524  -3.0344 -17.0379  12.4057  11.4665
#>   -8.1872   1.9474  -9.8392   1.3176   7.8220  20.5106  14.0397 -14.5463
#>    0.1881   1.7193  14.2382 -10.2011  -5.4146  -0.7746  10.6832  13.2266
#>   -4.4997   3.3831  -2.5354  15.3482  -3.1523   4.0547   1.6621  -7.5956
#>   -5.8841  10.5064  -3.8984  -1.8087  -2.2437   0.3013   5.5790   2.3654
#>   -0.8153   1.5495 -10.6338   9.1946   3.6116   2.6303  -6.7053 -14.7328
#>   -7.3348   5.1316  -3.4930   2.6760 -10.7297  -6.6452  -0.5947   1.1275
#>    1.5409  -5.8680   5.3638  -7.6750  -2.9600  -5.5917 -11.7459  14.5097
#>    0.1389  -9.9519   5.3985   6.7186 -10.4077  -4.7007  -6.5868   7.8764
#>  -15.2418  -0.1042  -9.0861   9.5038   6.4574 -16.6156  -3.5185  -7.3921
#>   -2.3448   0.8601 -11.1257 -14.8451  -7.6981   0.4425   7.6180   4.3863
#>    2.8902   9.4786  -5.8244 -14.0112   1.4965  17.5346  12.4112  -4.8505
#>   -1.0824   0.2375   7.1367  -9.4417  -7.3610  -9.4357   3.9947  17.6185
#>   -5.9409 -14.1050   2.3842   1.4447   6.5479  -2.9713  -5.6965  13.5628
#>    5.8684  -6.6025  -3.0483  -0.9700   0.6068  -9.5278 -15.8231  14.4439
#>    8.6863  -9.6832  -4.7707   5.1966  24.9754  13.4313   6.8675   2.7134
#>   -1.5309   9.2259   1.2896  -2.7691  -6.1147  -2.3758  29.8899  -4.5719
#>   -2.1610  -1.0335  -2.0070  -5.8368  -6.8901 -14.7272  24.6459  19.9214
#>    0.3874   4.7898  -2.7167   3.5264   0.7172   8.0933   1.5732  -3.4794
#>   -5.9317  -1.0218  -8.8980   9.2442  11.7047 -16.4275   5.2517   7.3198
#>  -13.1357  -1.4606   3.2042   6.7185 -11.9659  -9.8067 -12.2051  -3.9598
#>    1.4163   6.9033   4.5817  -2.8981   8.2961   5.2929   8.9960 -14.3995
#>   -4.1839   3.6118 -10.0947  -5.6397  -9.2227  -4.9447 -15.4553  -4.7722
#>    5.8583  -4.6736  -0.6184   1.0773  -1.4798  -7.4232   8.7624   0.9875
#>    5.7842  -9.0784   7.9809  12.5412  10.8854   3.0122  -8.4980   0.0674
#>   -1.6427  -5.8229   8.8754   6.4386 -10.4846   9.7098   8.7804  14.3993
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]