Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -6.4071e+00  2.6078e+00  3.9483e+00 -1.2661e+01  8.3329e+00 -3.1394e+00
#>  -2.4345e+00  3.1533e+00 -7.2896e+00 -1.3849e+01  7.5519e+00 -7.7644e-01
#>  -5.9447e-01  5.2631e+00 -1.8292e+00  1.2924e+01 -3.7410e+00 -1.6689e+01
#>   3.1497e+00 -6.2464e-01  9.2231e+00 -5.0403e-02 -2.3468e+00 -2.3400e+00
#>  -1.9102e+00 -1.7161e+00  2.1927e+00 -7.3535e+00 -1.3018e+01 -8.8460e+00
#>  -1.7552e+00 -7.8262e-01  1.3485e+01 -4.9994e+00  1.8360e+01  3.5977e+00
#>  -3.1188e+00  4.5201e+00 -6.9453e+00 -2.4324e+00  4.9565e-01  6.4190e+00
#>   3.5993e+00  1.1770e+00 -9.8476e+00 -8.6216e+00  2.5795e+00 -3.5503e+00
#>   7.8699e+00  3.9078e+00 -4.8337e-01  6.3239e+00 -2.1292e+00 -1.8932e+01
#>   4.3706e+00 -6.6494e+00 -2.5815e+00  4.1909e-01 -2.2742e+01  1.6795e+00
#>   1.0733e+00  4.4992e+00 -6.7757e+00 -2.9823e+00 -5.6723e-01 -5.0987e+00
#>   1.2551e+00  1.7194e+00 -6.2290e+00  2.2110e+00  5.0563e+00 -5.4199e+00
#>  -4.6261e-01 -4.1523e+00 -1.0478e+01 -5.5678e+00  1.4848e+00  1.2261e+01
#>   3.8711e+00 -3.2776e+00 -1.6493e+00 -8.4477e+00  1.8778e+01  1.5773e+01
#>   4.5990e+00 -5.1596e+00  1.8522e+00 -8.8443e+00 -1.0948e+01 -2.6477e+00
#>  -7.4680e+00 -1.1115e+01 -7.7143e-01  1.7383e+01 -1.4028e+01  5.6365e+00
#>   5.6044e+00 -4.3586e+00  1.7447e+00  8.8493e+00 -3.4417e+00  9.8761e+00
#>  -1.0709e+00 -3.5479e+00  2.3604e-01  1.2565e+01 -8.5238e+00 -1.3854e+01
#>  -3.6063e+00 -7.0519e+00  7.4996e+00  7.5295e+00  1.2366e+01  9.2184e+00
#>   1.8024e+00 -4.5135e+00  6.7474e+00 -3.0332e+00 -6.1617e-02 -2.2586e+01
#>  -4.7429e-01  1.0879e+01 -1.3628e+01  1.1050e+01  6.8359e+00 -4.3639e+00
#>   3.3412e+00  3.3371e+00 -3.3729e+00 -1.0502e+00  1.1882e+01 -6.7057e+00
#>   7.1257e-01 -9.2694e+00  3.9885e+00 -1.0531e+01  5.8701e+00 -1.0679e+01
#>  -7.5482e-01 -1.0913e+00  3.2494e+00  7.7903e+00  1.8185e+00  6.5341e+00
#>  -9.9860e-01  7.6774e+00 -3.6465e+00 -4.4149e+00  2.5838e+01 -1.2282e+00
#>   5.4031e+00 -4.0573e+00 -6.5159e+00 -2.4382e+00 -5.0274e+00 -3.3674e+00
#>  -5.0394e+00 -1.4548e+00 -3.7831e+00  1.2633e+01 -3.7233e+00 -4.1376e+00
#>   5.6833e+00  1.6311e+01  3.4953e+00 -1.9267e+00 -5.4100e+00  3.7394e+00
#>  -6.3776e+00  5.2697e+00  7.8855e-01  6.7621e+00 -2.3134e+01  1.7984e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]