Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.2860  -8.2412  -7.9485  -8.3612   1.6968   7.5072  10.7141  -5.4097
#>    9.1036  -3.1049  -4.3922 -12.4788  -6.3059  -7.0698   7.5480  -5.5043
#>    0.4862  -7.1239   6.2476   4.7290  -8.3241   3.6598  -2.6843 -10.0256
#>   -1.6193   2.7268  -4.5463   0.3818  13.1324  -9.4360  -5.5696   0.8601
#>   -3.1769   4.5111  -3.7869  18.1686  11.4397  -3.8763  -2.9049  -5.3827
#>    0.0365  -3.4618   7.0883  -1.1368   2.0479   0.1132 -13.5120  10.1992
#>    1.9196  14.8744 -17.9645  -7.0668  -9.7245   3.9010 -14.5326   2.3449
#>   -3.8945   6.8794  -3.7513  -0.2483  -4.5099   3.7440  10.6700   6.9788
#>    0.4926   4.5780   7.7533  -5.9905   4.7593   0.5294   1.3573 -13.9791
#>    2.9466   8.7477   0.9101   6.2421 -19.4981  -6.1718  11.5847  -3.9690
#>   -0.5024  -2.8891   7.3789   1.4415   0.9079   0.0294  -3.1463  -8.7915
#>    7.7537  -2.1809   5.1152  -6.3203   5.8987 -11.0847   3.9520   0.4797
#>   -5.0316  12.1341  -1.5793   2.7613  -4.7525  -8.3916   1.4442 -11.1461
#>   -1.3491   3.6074   4.2370   4.9514  -5.7839  -3.0986  -7.1204   7.5367
#>   -0.3880   3.5660  -3.1703 -10.8721  -2.1484  -9.1559   6.2537  -2.1818
#>    3.4657  -0.5636   1.0708   2.6112  -8.6691   2.8232  -8.8290  -0.7676
#>    2.0429   0.7525  11.0128  -0.5590  13.7204  -4.2518  -7.6945  -6.2459
#>    2.2557  -5.4753 -10.8301  13.5371   4.8284   4.3625  -0.6614 -11.1092
#>   -0.7546   9.7837 -13.8150   5.1116  -2.4770 -15.0876   2.4145  -4.5298
#>    7.5383  -3.0436  -9.1328  -1.3427   9.5708   5.1984  -6.8276   5.2839
#>   -3.2420   1.1857  -6.2845  -3.3598  19.3237   9.3933   7.9211  -0.8380
#>   -2.0926   2.4076  -5.0973 -16.5771  -0.9568   1.7999   4.1803  -2.8854
#>   -1.4407  -0.0847   6.3521   0.9302  12.2516  10.0750  -0.4201  -9.0320
#>   -2.0241  -1.7168   7.0896  -2.7706  -7.7671  -0.7875   2.7939   1.3030
#>   -5.2655   8.8263   5.1739   0.9144   3.2262 -10.9039  -9.5019  -6.8146
#>   -4.0750   5.0264   1.9645  15.0301 -13.9179  -5.1159 -15.8781   1.6018
#>    3.6109   3.2778  -5.8660   0.1393 -21.5992   5.2508 -13.4288   5.1918
#>    2.4559  -6.5593  -2.1429  -4.2050   9.9793   4.3323  15.9500   0.6105
#>   -9.0240 -10.1278  12.0153  23.8634  -0.3168   7.2277  -2.1068  -1.9479
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]