Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.3892  -1.2655  -5.5571  -3.5054  -5.8901   9.9016  10.8800  -4.7574
#>   -0.5959   4.3143   3.2347   4.0108  -2.6711 -11.4559   9.9474 -12.0776
#>   -2.2935  -4.8974   4.0048  14.1067   5.2709  17.0397   9.2945  -5.0231
#>    5.2421  -5.4113   1.7725   1.5902  -9.4696   6.7590  -1.1995  14.1310
#>   -0.2608   0.5138  -1.0287   3.8741   5.9497  20.5819  11.1851  -4.5996
#>   -4.5614   4.0606   4.7586   3.4665  -3.2711  -7.6347  12.0452  21.2837
#>   -1.6446  -7.9856  -3.8748   3.9272   8.7978  -3.9549   7.5227  -5.1635
#>    1.2704  -4.2429   0.6821  -0.1625  -0.2882   9.2622 -14.3686  -5.1654
#>   -7.5311  -0.7293  -2.4049   2.7814  -8.3950   8.0510  -6.6379   4.6024
#>    0.1962   9.0987  -0.4637  -1.5495  10.1976   8.4572  -0.9274  11.9754
#>   -2.5536  -6.6476   0.4638   8.1920  -3.0184  -7.4689   2.7098 -21.9453
#>   -2.4207   6.5072   0.8005  -4.9318  -0.1855   3.5021   2.9780   9.5801
#>   -0.2637   6.8558  -3.9572   4.7361  11.3233   0.1601   9.2122   8.6685
#>    4.9212   5.2407 -13.9340  -9.7146   0.5570  -1.4540  -0.4303  -6.5177
#>    6.8056  -2.8267  -7.7551  11.0489 -11.2319  -7.6398  -0.6906  -8.6370
#>    0.3238  -5.7361  10.0929  -4.0301   3.6270  -2.1023   3.9426   6.9516
#>   -3.4323  -0.3826  -4.8810   1.0093   9.6079 -23.0632 -15.7347   0.2113
#>   -0.4315  -4.1877   7.9772  -8.5627  13.2283  -2.7433   6.6950  -1.3294
#>   -2.0067  -8.2118   1.9324   4.9903   4.1500   3.7424   0.0460  14.6861
#>    3.0032   6.9347 -16.2576   1.3182   1.6738  -0.9324 -10.1411   6.1889
#>   -0.0669  12.0099  -0.0828   2.9903   4.1610  10.7590   0.4615   5.1154
#>   -1.0074  -5.4616   6.4077   4.0487  -6.4296  -3.1251  -3.4445  14.6759
#>   -1.5663  -4.0244   2.8914   6.4340   9.3128  -1.3365   6.1489  -9.0329
#>   -4.8355  -6.6911   8.3705  -1.1462 -12.4855  -3.1214  -9.9194   6.2613
#>    1.0595   1.4773  -9.3731   2.3321   3.9131  -8.3035   9.2170   4.4699
#>    1.4392  -1.5397   1.4312  12.8181  11.6449   7.6496  14.1417  -8.5951
#>   -3.9467  -2.8286   1.9408  -1.8934  -9.3549   3.5717   6.7457   8.0109
#>    1.7378  -4.4238   0.3325   4.3368  -4.8384  -9.8743   3.1372 -13.6169
#>    7.1329   0.2008   4.3880  -3.4388   1.3737 -25.2348   0.4898 -19.1167
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]