Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.0858  14.2445   3.7320  -9.6977  -3.0891   3.2190 -11.5927 -13.1485
#>    3.7839   6.7834  -4.7652  19.1040  11.0800   2.2563  -7.2731   0.4224
#>   -1.9693  -2.6184   9.4197   8.9845   7.5255  14.3273 -10.9132   4.0708
#>   -0.7165   7.3232   0.2280   6.5141  -7.0404   8.0311   9.3352   5.1523
#>   -6.0651  -0.5676  12.9432  -6.3944   2.4799   5.8297   5.3753  13.8977
#>   -3.3149  -0.9639   5.9795  10.3891  -3.9532  -3.8182   8.2069  14.9037
#>   -1.0288  -5.9669  -0.4652 -11.7812  12.0016  -6.2277   3.1283   3.3512
#>    4.6022   1.6859  -7.4474   3.2459 -10.3265 -10.5519  -4.3387  -8.0074
#>   -2.2490  -8.0172 -15.4810   5.0034 -20.7488  -2.1523   3.3626 -11.4189
#>    4.9873   2.4718 -11.4611  -0.5818 -11.0935  11.5660  -8.7914 -15.2595
#>   -3.1856   2.5598   3.1180   4.4562 -16.0319  18.5546  12.3115  12.0643
#>   -1.8413   1.1347  -1.3317  -3.8366  -7.7026  -4.3740  11.2569   1.2863
#>    8.3937   2.0429   1.6411   2.2330   4.9863  23.7365   6.7152  -1.5901
#>    8.7558   6.7101  -5.2226   7.7317   3.6074  -9.1110   6.9409  -6.9164
#>    1.8582   3.3143   1.4515  -0.6834 -17.5235   2.9819  -4.9979   1.8678
#>   -5.7522   2.3948 -21.6247  11.6102   1.7837  -9.2772  -1.2763  -3.7403
#>    3.7172  -1.0654   0.5219   1.6965  -4.5280  -3.0862  14.1352   4.5590
#>   -1.7775   1.7987  -4.7300  12.8717   0.3917  -0.6767 -13.3489  18.8117
#>   -6.4041   7.1143   0.3089   0.8113 -29.4457  -2.8107 -11.9302  10.0055
#>   -1.9874   4.0460  -8.7154   9.8292  -4.4754   6.4573  -8.9851  14.5994
#>    1.5870  -7.0139   4.4090  -1.8687  -2.7332  -4.7309   3.5839  -3.2266
#>    5.8192  12.3572   6.0990   1.2322  18.4434   5.4392   9.9870  -1.5738
#>    6.5712  -0.4337   1.0475  -5.6231   4.2148  -6.5293  13.7060   8.9049
#>   -2.9245  -4.3343   0.5074  -4.2653  14.3633  10.6977   5.5766  -1.4297
#>   -1.3898   1.9404   1.7856  -2.5412 -11.4315   9.4646  11.2138 -16.5461
#>   -0.3902  -8.6744   0.9302   1.8303 -23.9405   9.9959   1.3270   2.2979
#>   -5.4400  16.5931   1.6117   8.0918 -10.4663  15.8832  -1.9304  -7.7383
#>   -4.9295   5.9092   0.6986 -13.1072  18.4095   6.0553  -1.0968   2.0240
#>    1.0709  11.0461  -8.0256  13.5187  18.9691  12.0018   1.9166   1.1869
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]