Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.9165  -1.9605  -6.0635   2.0391  -0.2097  -1.6921   6.4260  -5.6514
#>   -9.8986  -2.0943  -5.9374 -11.8414   5.0924  18.2082  14.5313   0.8376
#>    0.4863  -8.6586   1.4312   8.6750  -7.3955   0.2181 -19.9656  -6.6311
#>   -1.9455  12.4815 -16.4634  -8.1866   0.9328  -3.7437  -2.9734  -3.6269
#>   -3.4327  -1.3190   3.5865   4.6871   4.9597   6.9372 -14.1669  10.1074
#>   -1.8798  -3.0863   4.0339 -15.2541 -13.2394  -8.1064  -7.6452  -5.8244
#>  -11.7783   4.4030   0.2591 -16.7475   1.9640  -0.3379  15.7403  29.7393
#>    9.0508 -11.0691   2.7199   3.0849   9.0816 -19.3493   0.4553  -4.1269
#>   -3.8026   7.7062  -5.9515   8.2725 -12.1027 -10.2202   9.7846  -0.6763
#>    0.7891   1.3435   0.6613  -7.7463   1.3483   4.3436  -9.4304  -8.5446
#>    6.3881  -7.1119 -14.1679  -5.3821  -4.1781  -5.5571  -5.0657   7.1675
#>    3.5246  -6.7289 -10.2356  12.1101   9.8271  10.4903  -3.2485   5.8349
#>   -3.2289  10.8551   6.5853   4.8749  -6.4374  -2.3472   2.7391  -0.3361
#>   -5.0932  -3.7021   5.4525  -2.6402   3.9482  -3.7662  16.0805  -2.6802
#>   -7.0819   3.1532   7.1519   0.9070  -1.7780  -2.8387  13.7283   5.7801
#>    7.9756  -2.5220  -0.0395  -4.4268 -12.8796 -20.4192 -21.6683   2.6041
#>   -1.7364  -3.6868  -5.9319 -16.2789   3.1991  -4.7192  -1.0052  -7.7572
#>    3.7659  -1.2020  -7.4081   6.0867   0.9308   0.6075 -13.0219  -2.4457
#>    1.2922  -0.3587  -3.6444   9.6340   6.8590  -5.0395  -4.1942  -1.1680
#>   -0.7145   5.2454  -5.9869  -4.4099  -3.5355   9.0387  -0.6612  -6.2132
#>   -2.9827  -1.7779  -8.2664  13.0086 -13.6019  -9.3680   1.3067  -5.8024
#>   -8.8068   5.7806  -0.7246 -11.2078 -14.8587   6.8903   2.6868 -11.2433
#>   -3.4727  -1.4408   1.4321  -7.1630  -6.1309  -7.8096  -5.8482  -2.7106
#>    5.4015  -0.0207  18.0308  17.8367  -7.6702  17.2021  -0.3811  -0.1347
#>    0.0904  -2.5642   5.8717   0.3850   7.7751   3.3639   3.4614  -7.8238
#>    2.1923   7.2061  -1.9068   5.6219  -9.1810   8.4984   2.9990  16.3472
#>   -1.4647   0.2195   4.9163   0.9739  -0.1632   2.9528   0.1646  -3.1028
#>    1.9969   2.0157 -12.6881   0.0729  -4.6237   1.8279   2.2481   0.3842
#>    1.8122  -0.9651 -13.4599  -0.8677   9.6499  17.1524  -2.4947  13.1024
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]