Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.9405  -6.3681  -2.7841  -1.9620  -1.5070  -3.2464  21.2780  -6.6988
#>   -0.9062  -2.0141  11.2753 -17.5365  15.8419   8.7718   0.3803   0.8163
#>    2.4309  -1.9109  -7.4893   6.3390 -14.7651  11.3350   1.4050  -7.5767
#>    0.6218   0.1351  -0.0149   7.1460   4.3531  -6.4096  -3.8826  -6.3550
#>   -1.7989  -2.2158   6.1010   2.8539 -11.4642  -0.4528  -5.0785 -17.5629
#>   -1.2630   4.4429  10.0300   2.7126  -5.3370  -9.4191   0.3280  -3.6111
#>    0.2730   0.2566  -4.8280   0.6307  16.7091   3.4005  16.6462  -9.7763
#>   -4.5594  11.3528   2.0107   4.2827   0.7081  18.7304  -4.3845  -4.6291
#>   -1.7226  -2.0580   8.7124  -8.7720  21.1422   6.1386   7.6615  -6.9854
#>    0.3267   0.3493  -7.3520  -8.3187  -5.1290  12.3192 -12.8989   8.8986
#>   -0.0867   8.1378 -10.8260  -5.2423  -1.2204  -0.1530  16.9738  -1.1233
#>   -3.1714   6.0016 -10.8740  12.0902  -7.4772  -9.1394  13.3509   2.9756
#>   -1.3045  -9.8345  13.5075   3.7107   1.2445   5.4304  10.7743   1.4952
#>   -0.7038  11.6791  -7.1530   2.9978  -1.5070   1.0599   7.4209 -12.2798
#>   -1.0680   1.5845   1.4836  -6.6674  -6.5887   0.1503   1.2972  13.0052
#>   -4.1155  -3.8678   9.0709   3.8873   0.8104 -10.7754  -1.4670  -1.0374
#>    1.3645  -3.7306  -3.2035  16.6586 -10.2788  -1.5338 -15.5122  18.2183
#>    3.1003  -5.0197   2.2527  11.3013 -12.2064   7.1300   2.8769  -6.5173
#>   -2.3289   0.3608   5.6106 -15.2124   4.9824  -7.4177  18.2567 -15.7722
#>    2.7249 -14.2865  17.0064  18.4728   6.1325  -2.5150  -6.6414  -7.1596
#>    4.0682   3.9407  -5.1257  13.6320 -11.2239  -2.9577   2.4284   5.0568
#>    0.4549  -8.9609   1.4715 -17.1802  10.4674  -5.5765   4.6347 -10.9331
#>    0.2455   6.5284   3.3312  -3.8795  -9.0110   1.0103   3.2870   3.0713
#>   -5.0591   7.0763   1.8605  -3.5043  10.1304   6.2095 -21.9282   0.2210
#>   -4.9375   4.6274   3.8444  -1.6431  -7.6252  -7.3193  -8.1512   4.8094
#>   -3.4329   2.2587   4.1103   6.1330  15.9285 -13.2259   4.0470   0.6550
#>    2.5166  -2.4644   1.3791  -2.6946   0.3867   1.2592 -16.1973  -5.4366
#>   -0.6509  -1.9547  -6.2264  -4.7928  -2.6476 -14.1432  21.2448  -4.5099
#>    0.9749  -3.3538  -4.6862  -6.5388   0.8184   4.1019  -3.8886  -0.6801
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]