Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.3525  -6.6108   2.9182  -3.7678   3.0222   1.6437  21.2450   0.5543
#>    2.5450  -0.5666   7.2284   2.5320   2.9874   0.5175   1.1394   5.5966
#>    4.6627   1.0603  -0.9642  -3.0353  -4.1522   1.7453  -4.3685  -8.2498
#>    4.4770  -2.1378   5.3843   0.5520  11.8367  -5.6246   3.4510  -1.1096
#>   -5.2510  -2.3560  13.6453  -0.5152  -7.3411  -8.5485 -15.1979 -13.0175
#>   -2.0504   2.9464  -3.9468   2.1960   1.9384   5.0757  -7.9436  -7.8793
#>    1.6626   0.4738   2.7441   9.2907   2.4693  -3.3334  -3.3334   7.7610
#>    7.7210   5.9457   5.4060   0.6990   8.8663  -9.6196   1.8705   3.1307
#>    3.6523   1.0551  10.4459  -7.4616   6.8308   0.5583   4.2376  -4.7036
#>    0.2198   1.5374  -8.0781   0.1237  -8.2428  32.1835 -11.2186   4.2665
#>   -2.5640   1.0586  -9.6012   3.7821  -3.7930   3.1748  -2.2220  -4.0835
#>   -0.5794   7.4185  -0.7110 -10.7246  -3.1377  -0.3901   3.7489  14.7796
#>   -3.9005   1.7609 -18.8719   0.0549  -8.9551  -3.0161   1.2763  10.7899
#>   -8.5702  10.2732  -6.0143   0.2165   3.9430  -2.5596   8.4738   3.8845
#>    5.2652   5.6590  -2.3087  -3.8267  -1.0348   9.1446   1.2218  -0.3344
#>    7.2345   0.8623   1.5139 -17.9294  -4.6083  -0.5560  -8.0735  -6.3841
#>   -6.9607  -7.7416  -8.1939  -6.1921   8.1163  -0.3838  -8.8553  -9.2982
#>   -2.5713  -0.2428  -0.5704   0.9722  19.0957   0.1735  -7.7789  -8.3274
#>    2.4039   9.0148  -4.7229   7.0492   7.4456  -7.8996  -2.8051  11.4967
#>    0.0073  -4.7650  -5.6390   7.0284   5.8084   8.1753   4.4604   5.9626
#>   -5.1764  -4.2100   0.5177 -13.8712   3.5698 -19.8131   3.9691  -0.8054
#>    5.2646   1.1268   1.4196  -1.8157   6.2732   7.0680   0.9220 -17.9295
#>    4.8114   2.4997   5.5538   7.9642   1.3354   6.3673  -9.8563 -15.5710
#>   -4.2916   0.6009 -10.5148   7.6383  10.7157   1.6761 -17.6388  -1.5188
#>    0.0449  -0.2263  10.8594   5.3367   8.3020  15.1928   2.7978  15.4336
#>   -0.8239   2.6277   7.6364   4.5510   4.6193 -11.9401  -6.2048  -4.3229
#>   -5.6499   2.4975 -13.3176   7.9740  -8.6966   3.7442   1.5481 -10.3399
#>   -3.8443   0.7125  -4.7980  -4.7742  -5.1713   3.6859  -4.7945  -4.3703
#>   -7.9941  -2.3136  -1.8630  -8.0887  -5.4606 -14.5242  -3.0549  -8.1486
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]