Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.4690 -12.2859   5.4184 -10.2508  -6.4970  -1.7288   6.0151  -9.3156
#>   -4.7871  -3.3558  -7.6629  14.0455  11.8408  -2.6509 -12.1947  -2.1286
#>    1.7116  -0.6976  11.9946  15.0105   5.9850  -5.5880 -11.8984   8.7102
#>    1.6920   4.9692  -0.2773  12.3288  -9.0793 -12.7480  -7.7709 -12.3840
#>   -8.1337   9.6896  10.0670   9.2124  -1.8927  -1.9253   5.0184  -2.0188
#>   -1.1757   8.5416  -0.0820   0.5023  -6.9403  17.2469   2.0129  12.6452
#>  -12.5353  -2.0307  -5.1089  -3.2426   6.7670  -2.6615  -0.6659  10.0685
#>    2.7188  11.3910   6.3797  -9.3060 -19.6522 -17.7753  10.3737   5.4835
#>    0.0552  -0.9103  -6.1036  -7.0815  -2.7746   7.4910   4.6038   2.1509
#>    3.4573  11.2847   6.5408  -0.0168   0.9296  -1.2698   6.2335  -2.6304
#>    6.6279   5.4135   5.7856  -3.2734  -7.6684  -2.0235   9.6148  -0.7612
#>    0.2973   3.7136  -2.6192  -7.4162  13.3978   5.3203  -0.3854  -2.6353
#>    5.8550   6.4055  -2.1144  -1.6349 -18.4373 -17.4849   3.1143  -7.5177
#>   -2.9873  -4.0711   4.1131  -0.5241 -15.0249 -17.5301   0.3618   8.5221
#>   -1.7762   3.3178   0.9837  -0.0237  15.6793   0.1246  15.0470  -1.2124
#>   -0.7485   4.5513  -3.4243  -1.2264   5.7048  -5.7819  -5.5398   5.8016
#>   -1.2626  -0.7815  11.7722  -9.9498  -5.2849  -8.5516   2.1418   7.4351
#>    5.1815   1.6862   0.0213   7.7245  20.0925  12.6287 -10.9776   1.2299
#>   -2.1925   3.7259  14.8811   2.9057   4.3640  -7.5960   2.5417 -11.0191
#>   -1.5485  -8.9514   3.5266  -6.5638 -13.9585  -5.8295  -5.3843   4.1846
#>   -2.3642   1.5388   2.8550   1.1645 -18.3024   0.2486  10.1491  -5.0395
#>    3.7070   8.8017   7.5180   3.5346  -9.6118 -24.5372  -5.6849  -8.0146
#>   -5.3518  -5.9632  -4.2972  -1.2932  15.2609  -2.6958   9.3308  -0.8419
#>   -0.6558  -8.6846  -5.1667  -0.6041  -1.2011   8.4247   5.1618  -3.3642
#>   -2.6049  -2.3232   4.5600 -17.4688   9.0317 -11.1757  -2.1048  -6.3327
#>    4.3312   6.5355   5.5175  -6.6749   0.4133 -13.5061  -8.2549  -0.1029
#>   -7.1773  -8.4789 -13.7413  22.1415   4.2101  29.9650  -6.5688   9.5494
#>    1.8120   1.0562  -4.4844  -4.2883 -13.8730  -1.8233  -2.7032 -14.2808
#>   -1.5508  -2.0325   1.4977   8.0910  15.4418   2.3369  -2.8394   1.7091
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]