Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -3.7376e+00 -1.9019e-01  3.1960e+00 -1.3850e+01  4.9379e+00 -9.4491e+00
#>   1.2751e-01 -1.0464e-01 -4.1557e+00  6.9542e+00  1.1205e+01 -2.1683e+00
#>  -1.5583e+00 -8.7668e+00 -1.4891e+00  1.6900e+01 -5.3088e+00  4.1437e-02
#>  -2.5284e+00 -1.2420e+00  1.2497e+00  8.0608e-01  3.6145e+00  1.0092e+00
#>   3.0315e+00 -7.2669e+00  6.7115e+00 -1.2993e+01  1.3748e+01 -2.0825e+01
#>   4.5335e+00 -2.3895e+00 -7.7162e+00 -1.0949e+01 -1.1072e+01 -2.0955e+01
#>  -3.0468e-01  1.1991e+01 -3.3932e+00  1.4787e+00  8.4571e+00  1.0752e+01
#>   4.0208e-01  1.3154e+01 -7.4612e+00 -8.2842e+00 -2.7509e+00  1.2063e+00
#>  -3.2393e+00  1.7853e+00 -5.4521e+00 -5.9363e+00 -7.6259e+00  2.2595e+00
#>  -9.3720e-01  6.7858e+00 -1.5873e-02 -2.6532e-01  1.0473e+01 -1.3472e+01
#>   4.4990e-01  6.3090e+00  1.0016e+01  4.6517e+00  5.9895e-01  7.2758e+00
#>  -2.8329e+00 -2.5952e+00 -7.6021e+00 -3.4221e-01 -1.4448e+01  3.6746e-01
#>  -7.9499e-01  1.0319e+01 -6.8386e+00  1.2747e+01 -3.8685e+00  1.5584e+01
#>   3.4329e+00  7.8186e+00 -1.1765e+00  1.0505e+01  1.2907e+01 -1.4393e+01
#>  -1.3491e+00 -1.1147e+00  6.7837e-02  7.6991e+00 -1.1491e+01 -1.0002e+01
#>  -4.2076e+00  4.3336e+00 -7.1856e+00 -9.8519e-01  1.0964e+00 -1.2805e+01
#>   1.7950e+00 -3.0016e+00 -3.1770e+00  9.2961e+00 -2.8715e+00  9.4054e-01
#>   4.8210e+00 -5.5693e+00  3.8027e+00  1.7533e+00  9.4805e+00  3.0613e+00
#>   1.8625e+00  1.2191e+00 -2.2060e+00  5.7353e+00  6.1759e+00  7.5235e-01
#>   4.7675e+00 -1.1304e+01  6.2839e+00  1.0687e+01 -4.4682e-01 -1.1035e+00
#>   1.6944e+00 -6.4609e+00 -1.1380e+01 -4.2229e+00 -1.7338e+01 -1.1693e+00
#>   3.0164e+00  1.1239e+00 -7.4541e+00 -1.4648e+00 -1.8958e+00  2.5677e+00
#>  -4.7171e+00  9.2808e+00  1.8634e+00  8.1043e+00 -1.1725e+00  9.8549e+00
#>   5.9584e+00 -4.9519e+00  3.2992e+00 -8.2923e+00  1.2331e+01 -5.5699e+00
#>   1.3431e+00  1.0105e+00 -1.3524e+00  1.0426e+01  1.1715e+00  1.4881e+01
#>  -4.7119e+00  2.5791e+00 -1.2499e+01  8.3061e+00  1.4975e+01  9.6673e-01
#>  -3.8733e+00 -4.0900e+00 -1.2471e+01 -1.1018e+01 -5.3390e+00 -1.4819e+01
#>   2.7910e+00 -2.6432e+00 -8.3053e-01 -3.4569e+00  6.3087e+00 -2.1119e+00
#>   1.6855e+00 -3.5572e+00  6.3362e+00  1.8724e+01 -7.9022e+00  2.8980e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]