Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.4545  -3.0841  10.9212  -6.9091  11.2319   9.2112  26.0166  11.3608
#>    1.0563   1.5781  -3.2733  -4.9468 -10.5697   4.1256  10.1429   6.6568
#>   -0.3195   2.7293  -7.3392   5.1367   0.3458   4.8081   6.1660  -2.0380
#>   -0.9809  -4.6159   2.5861   3.5201   7.0862  -0.5602  -7.4277  -3.8495
#>    2.4207  -1.1704  12.8165  -3.3105   0.3445  -2.4869 -26.7632   6.9597
#>   -1.1356  -7.1499  -6.0993   1.5334   5.2890  11.2988  -2.1097   2.4677
#>   -0.8997 -11.5668  -7.9864  -8.3144  20.9966  -0.3245   5.6410  -4.3587
#>   -0.1912  -4.2979   1.9536   5.0328  -1.3287  10.1586  13.7901  -0.4882
#>    1.6258   3.5228   1.4941  -2.9773   4.3013  -0.9727 -11.8422  -5.3645
#>    1.0461   0.6298 -13.8540  -3.5552  -6.7233  -4.5290  18.6666 -12.1252
#>    1.3079  -4.4703  -7.8092 -11.5833   0.8652   0.7643   8.9190  -2.3385
#>    0.1523   1.7102   1.9087   4.2549  -8.5495  -6.8162  -6.7570   2.1562
#>    0.5159   0.5184  -3.6228  -0.7518   2.4330 -10.0361  -3.3525  -3.4615
#>    0.2078   3.9434 -12.7837  -4.2684  -9.0487  -6.0372   0.9225   3.4081
#>    1.2529  10.5951   5.4151 -10.5414  -3.1614  10.2409   0.8278   4.8645
#>    0.1070  -2.1410  -0.5082  10.8746  12.2159  11.4970  11.2446  11.1114
#>    2.9878  -4.0384  -9.7463 -10.0068   3.3934   4.9324  -2.8141   0.1049
#>   -0.4394   5.1426   3.4075   1.2386 -16.1679  -8.8612 -10.4683   6.2179
#>   -1.8140  -0.0430  -2.3656   7.4646   5.2275  20.4027   5.9944   3.1223
#>   -2.1639  -6.2283   7.2004  20.8030  19.3697  -5.4775  -1.8311  -7.7904
#>    3.3277   6.5751  -3.4308  -5.8520 -13.0916  -6.0465   4.8248  14.8817
#>   -2.7635  -2.5451   0.6410   4.8167  -0.9856  -4.2302   8.3732  -3.7541
#>    3.7191  -2.7242  -9.1673   0.1718  -2.7399  -6.5768  12.1954  13.9388
#>    1.5111  -3.4055  -8.2621  -7.8137  -9.8181  -6.5047  -6.8562 -19.0570
#>   -5.1274  -3.5471   2.6353   7.1153  -3.0583  -8.3070 -11.6537 -22.3198
#>   -1.3458   3.2083  -5.1778   6.5216  11.1358 -11.7478   3.9666   6.9582
#>   -2.6804  -1.3531   5.0435  10.6087   4.0857  -3.3703 -15.9045  -5.4902
#>    7.7035  -5.7118  -2.1861 -15.8644  -2.7140 -12.2170  11.6805  -4.6327
#>    0.2281  -5.4612  -8.3300   5.9950  -5.6787   5.7459  -7.7259   6.6994
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]