Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.2662  -4.7383   6.6909  -4.3401  -8.3000  -8.3167  -8.6570  -6.2922
#>   -0.0058  -2.1172  -5.3548  12.0814   3.6363   2.5692   2.7172   1.7470
#>   -2.2487  -4.1902  14.5500  -3.0441   2.1697  -8.0712   9.0821   2.2446
#>    0.2965   4.7182  12.1914   4.0225  -6.1169   7.1535  -1.1740  18.2122
#>    0.9878  -0.4347   2.4261  11.9114  -4.1394   1.3579  -1.9127   8.6859
#>   -4.4468  -3.2498   9.6956  -0.3422  10.6258  -0.3171  -1.8712  12.2695
#>    2.9730   5.8971  -1.7098   8.7611   0.9938  -6.1677  -2.4107   3.4882
#>    0.9135   2.4801   5.5986  -0.5619  -6.2841  13.4896  15.1016  -4.0237
#>    6.4177   2.1826  -7.2870   5.8987   5.3500   1.3353   1.4614   4.8246
#>    8.2281   0.6204  -3.8600  -0.9603  -1.9478   9.8504   4.5190  -6.2989
#>   -1.5770  -2.9421  -1.4781   6.2809  -1.4692  -1.6985   5.0541  -8.5901
#>   -0.9034   2.2222  11.4066 -10.4945  13.1221  -6.7360   8.9007   7.8575
#>    1.5624  -9.3689  -5.1828  -0.4661   7.6957  13.3039  -8.0232  -2.8615
#>   -2.4211  -0.0811  -2.2486  -5.3551  -0.5474   1.7431  13.9860  -3.1789
#>   -1.1878  -3.9144   4.1063  -2.0578  10.2379  12.0047  -3.1621   5.4305
#>    7.5838  -1.4001  -0.9353   9.4345  -8.2316  11.6727   1.5706 -10.5466
#>   -2.7446  -8.4286  -2.6500  -6.8376   2.8040   9.0955  14.0497  -3.3808
#>   -0.9165   4.9515 -11.5597  -2.7566  13.7560   7.4701   2.6630   0.1138
#>   -0.0384  -8.2400   0.7966   1.7543  15.7253 -14.3861  -5.2557   0.7640
#>   -4.4335   9.8924   0.9374   2.8709  18.9793  -1.1562   8.1827   3.3669
#>   -3.9227   2.4450   7.0980   7.3591  10.2121   0.2616  -5.1990  -0.5170
#>    3.8378  -2.1162  -0.9140   3.8898  18.0769  -0.0331  13.8009   2.7759
#>   -1.6956  -1.1332  -8.7078 -13.9119   4.2374  -4.0996   1.5015   2.9264
#>    2.1728  -5.0447 -12.4625   1.5557   8.1602   4.8860  -5.8266   7.1732
#>   -1.7321  -1.2952   5.5441  -6.2089 -15.0487   7.2370  -7.6732  -1.1450
#>   -0.5694   0.1714  -6.7418  12.4419  -7.8856  -0.9313   5.1471 -13.4560
#>   -3.5250  -3.2236  18.0255   1.9828 -20.4529  -6.2042   1.8976   7.2297
#>   -6.9511  -2.6304  11.1359   3.9570   7.5189  -1.2562   2.2734  -8.4074
#>   -6.9131  -2.6720   3.0027 -18.4241   8.1369  -7.8142 -18.6697  -3.8561
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]