Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.1968  -1.1781   0.5593  18.1018  -6.4119 -14.0689  -0.8608   8.7814
#>    6.8240  -1.0493  -3.7310  -9.9747   0.7463   1.8993 -13.8479   8.3981
#>    0.7868   5.7192   0.2656   8.2467  -5.5299  -3.3086   8.7812  -5.1039
#>    1.8897  -7.6595   5.1682   7.7626  -8.2805 -20.7452   7.1887   9.6588
#>   -1.3882  -3.3224   0.6438  -6.7082  13.1499   4.9221 -13.6098   0.1079
#>   -2.5650  -4.5075  11.3554  -6.7632  -7.3580  -0.9879  31.5211   4.4815
#>    1.7670   8.6316  -4.3292   2.3597   4.7826 -11.0710  -6.2587  15.5663
#>   -3.9067  13.4637  -9.0234  11.5266  -3.5927  -5.5533  18.7344   3.1099
#>   -4.8203  -4.2432   2.5085  -8.2547  -6.6200  -5.1996 -13.8433   3.9179
#>   -0.1232   6.4336 -17.1555  -0.0413   4.0761  14.9997  -7.9578 -14.9351
#>   -2.9704  -2.2097  -4.5978  14.1271  -2.8964   2.7143   1.9537   0.1207
#>    5.0781  -3.7312  -6.6561  -0.8973   3.0890  -4.5424   3.6497  12.9736
#>   -3.1416  -7.4369  16.9109  -5.5328  -4.5203 -11.0226  10.9225 -10.8181
#>   -0.0923   7.5447 -16.3242  10.6636  -2.5622  -3.7584  13.4727  -0.4182
#>    0.2502   5.9232 -12.7295 -12.7506   9.6848   1.0284  -4.2534  -7.3628
#>   -2.6075   5.3847   7.3399   5.1413  16.0196  13.2570  -1.9693  12.6292
#>   -0.8876  -5.3881  -6.2228  -0.1708   8.8201  14.3326  -7.9516  18.6107
#>    0.3712   8.1021  -1.5966  16.4730  -0.6532   9.4608 -13.9541   4.6890
#>    5.9956  -5.8010   5.1560 -14.4877   0.2634  14.4082   2.9284  13.3137
#>    3.1885 -12.4787  23.5562  -2.8534   0.0031  -9.0513   0.8608   2.1775
#>    1.1106  -6.2020   1.8479  -1.3254  23.8714 -18.1652   2.6845  -2.7962
#>    4.3228  -0.2025   3.7801  -5.2589  -1.1260   3.0095   6.1564  14.8102
#>    0.4550   0.2368   0.1508   0.0034  -9.8983  -9.1817 -18.9778  -3.2830
#>    6.1856   0.1834  -4.1015  -3.4589   0.7410  -6.6150  -8.8596  -3.1756
#>   -3.7443   7.5167   3.6472   2.3622   1.4488 -16.6450   2.7921  -1.7260
#>   -0.7674  -0.5558   1.0347 -18.9649   2.7952  -8.0577   5.8548  -2.1372
#>    2.4998  -4.6026   3.9500   7.9020  12.7096   3.5702  -5.1108  11.4035
#>   -7.8477   7.9533  -0.8689   2.0493 -10.9877   0.8587 -10.6387  -8.6853
#>    2.7077  -4.7882 -16.6785   3.4881   5.0912   6.3058   3.8498  -7.8654
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]