Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  5.1534e+00  1.1324e+01  5.6835e+00  8.2419e-01  1.3460e+01 -6.2870e-01
#>   1.0914e+00 -8.8884e+00 -9.4304e+00 -9.3187e+00  6.4531e-01 -5.5888e+00
#>  -1.1326e-01  7.0088e+00  9.1224e+00 -1.3071e+01 -1.6836e+01 -1.0400e+01
#>   1.8541e-02  8.6375e+00  1.8009e+01 -4.3005e+00 -9.4762e-01 -1.1969e+01
#>  -1.2366e+01  9.5995e+00 -6.6475e+00 -3.7820e-01  4.9379e+00  1.9457e+01
#>  -3.6156e+00 -3.3818e+00 -8.5999e+00  2.8418e+00  3.9353e-01 -8.3156e+00
#>   2.4988e+00 -8.1607e+00  1.5799e+01  1.5732e+01  6.9089e+00  1.4797e+01
#>  -5.2076e+00 -6.1044e+00 -1.7480e+00  1.5348e+01  2.0097e+01  7.2626e+00
#>  -1.7138e+00  7.6028e+00  4.1749e+00 -1.1373e+01  4.2741e+00  5.1843e+00
#>   3.1000e+00 -2.0715e+00  3.2221e+00 -1.5533e+01  4.4582e-01 -3.7792e+00
#>   4.5704e-01  1.3275e+00 -8.1047e+00  2.4562e+00 -8.2596e+00  4.7879e+00
#>   1.9395e+00 -4.9894e-01 -2.4141e+00  7.7514e+00  8.7180e+00  1.3251e+01
#>   2.3839e+00 -3.3623e+00 -1.0577e+01  1.7206e+01 -1.4354e+01 -5.3282e+00
#>   2.6848e+00 -6.7054e-01  6.4804e+00  9.6905e+00 -1.8353e+00  3.6589e+00
#>   3.7712e-01 -2.9340e+00  6.9123e+00 -5.4707e+00 -1.8718e+01 -1.0038e+01
#>   3.7279e+00  5.9914e-02  1.8463e+00 -2.0146e+01 -1.5509e+01  3.5888e+00
#>  -8.3151e+00 -3.6507e+00  8.0869e-01 -1.9076e+01  1.1202e+01  1.5130e+01
#>   2.6183e+00  7.5104e+00 -5.3128e+00  1.9247e+00  3.9304e+00  1.7381e+00
#>  -4.7970e+00  2.1571e+00  4.7012e+00  9.8443e+00 -9.7777e+00  1.5801e-01
#>   3.3941e-01  3.7320e+00 -1.7270e+01  8.1131e+00 -1.0771e+01  1.4539e+00
#>  -1.5943e+00 -6.1894e+00  1.7393e+01 -1.7166e+00  2.6647e+00  1.3318e+01
#>  -1.2164e+01 -5.9861e-01  6.4719e+00 -5.6280e+00  4.5833e+00  1.5868e+01
#>  -8.2742e+00 -7.8774e+00 -2.7503e+00 -1.1751e+01 -4.9525e+00 -1.2141e+00
#>   3.3554e+00  1.3091e+01 -2.7750e+00 -1.1090e+01  1.5128e+01  2.8952e+00
#>   8.2672e+00  1.5545e+01  7.7769e+00 -1.3049e-01 -6.5962e+00  1.2863e+00
#>  -5.7772e+00  4.8582e-01  8.6441e+00 -4.1224e+00 -1.9219e+01  3.9047e+00
#>  -2.5879e+00 -1.1699e+01 -7.1774e+00  5.7168e+00  2.8975e+00 -2.0836e+01
#>   1.6472e+00  1.1743e+01 -4.2896e+00 -9.1453e+00  2.0708e+01  7.0145e+00
#>  -6.3913e+00 -4.1615e+00  1.1771e+01 -7.3382e+00  6.0398e+00  2.9464e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]