Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.8047  -2.6207   5.7633  13.3039  -4.0373  12.2551  -7.3413 -11.2976
#>    2.6096  -0.1024   2.6316  -6.3285  -3.2360   5.6601   2.1933  12.0639
#>    1.5961  -0.1171   0.8681  -1.7999   3.8697   3.5652  10.9547  -9.8713
#>    2.9794  -3.6050  -7.4980 -10.0766  19.2843  12.4435   2.1382  16.1470
#>   -0.3711  -4.0863  10.8270 -15.0806   9.3351   7.4108   4.6173   1.4591
#>   -2.1784  -7.4240  15.5783   6.5194   3.5944  -3.6951   1.5746   0.0820
#>   -3.9361  -7.3868   4.5942  -0.6890  -9.3278  -4.9997   4.8561  19.1913
#>   -1.5633   0.1429   0.9636  -1.2722 -18.1813  11.7843   6.8601 -22.5307
#>    7.8022  -1.4867 -13.7986   4.5614  -3.1620   4.8740   7.6869   3.3688
#>    3.2048  -9.5408  -8.0936  19.6564  -2.7941  11.9933   7.0701 -28.0072
#>   -2.0741   0.9349   5.4943   0.9087  -4.5853  -0.9088  -6.6539   5.5500
#>   -3.6006  -0.1049  -1.9923  13.7405  -5.8785  -2.8199  -3.4413   7.6612
#>    4.1098  -2.9058 -14.2701  12.0800  17.0652  -6.1976   2.2618   6.9915
#>   -0.0920  -4.9722   2.0921 -15.3755  17.3891   0.1948 -11.1514  -9.0474
#>   -3.5357   7.3730   4.8430   2.0695   0.0132  -2.8749  -1.6219 -11.8286
#>   -3.5903  -5.9056  -2.1158  20.4538   0.7726 -15.5031 -10.6137 -15.6691
#>    4.4591  -8.9946   2.1682  -5.1618  -4.0660   0.3873  12.3416  22.5118
#>   -4.9237   4.2936  -1.5132  -1.7673  10.4786  10.3201   7.8714  -4.1920
#>   -1.1305   8.5203  -3.9817  -2.1440 -12.3824 -19.6851  12.0374  16.7455
#>   -4.2279   0.7524  -0.7074  -1.4814 -22.2824  -0.0702  24.3189  -0.2784
#>    1.9868   6.8079   7.1363  -7.4898 -12.9580  -0.4718  -4.5530  13.8853
#>   -5.8421  11.8088  -6.3895   1.9566  -8.4891 -12.2354  14.0728  -2.3782
#>    1.8134   3.4250  -8.3145  16.5459   5.5404   6.6642  -5.7249 -13.7320
#>    5.3022  -0.3911  -9.7293   4.6165  -4.7555  -3.5539 -10.3060  14.3547
#>    0.6311   2.9731  10.5016  -9.3779 -10.7883   2.9377   7.3078  -2.0266
#>   -0.5939  -5.6225   4.1452   3.2519 -16.6833  -4.8651  -1.3980  22.3373
#>   -5.6136   4.6908  12.6139 -13.6881 -22.3507  -5.1841   6.5983  24.9246
#>   -1.8884   4.8789   2.2877  -3.1645   6.3985  -7.2532  18.2137 -18.8673
#>    2.0769   2.8879   0.5638  -2.5591   7.9662   7.5436  -7.5035  -5.2108
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]