Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.0239   3.4971   5.3005   0.5712 -20.7807  -1.7884  -1.4054  12.8170
#>    1.0910   4.5249   0.6188  -2.8707   9.9205  13.1556 -14.2843   5.4120
#>   -0.0308  -2.9664  -2.1950  13.4585 -11.2888   8.2897  -4.8369   4.3867
#>   -0.5277  -0.7674   4.3655  -8.8051 -15.2030   5.9056   2.6404 -19.5768
#>   -4.6452  -5.1672   8.6633   1.3374  -1.9302   0.9336  -7.1486   7.3941
#>    6.1942   2.6857   2.3660   2.3369   1.3332  15.2682  13.5917   4.8034
#>    9.0763   1.5853  -4.7432  -2.0021   2.4036  -0.0219  -8.0342  16.8202
#>    2.8000   1.2213   5.6711  -4.4966   7.4176   2.5404   3.7015  -2.8946
#>    6.6464  -0.0926   2.4394   3.5219   1.6760   1.1146  16.0893   2.2408
#>   -1.8883  -4.2949  -2.1421   5.4569  10.0587   1.2529   2.9043   5.1401
#>   -2.9720  -6.6145   5.6299  -3.3044  -1.3197  -2.5373  25.0878  -0.4071
#>    0.7057   0.9090  -2.1720   6.8426  -2.5947   0.2253  -0.0426   1.6921
#>    0.8244  -6.2255  -3.3802  -3.8741  14.4263  -0.9367  -2.7656 -17.9032
#>   -0.0603   2.2796   5.2659  -3.6099  -8.4768  -1.6956 -14.7414   5.7479
#>   -0.6270  -2.5957  -2.4319   0.2267 -14.4605  -8.5468   0.4458  -4.0189
#>    4.3138   2.5877  -3.5777   5.6031   2.1936  10.6087  -5.7681 -11.2234
#>    0.9896   0.7667  -0.7593  11.2077   4.9184 -10.4591 -13.1013  10.4338
#>   -0.3282  -4.7967   0.6283 -10.3178  -5.7847  -8.7489 -16.4949   6.4246
#>   -4.6341  -2.9146  -1.7163   5.0995  12.6367  -3.8060   3.4582  -1.7085
#>    3.5312  -1.9064   5.1924 -10.1903  -6.1121   3.0053 -11.5933  -1.0670
#>    1.0039   8.6635  -4.1790  -9.1668  -6.6300   0.6642   1.6878 -20.9059
#>   -4.8636   7.4619  -5.8673  -8.8865  -6.1997   3.0107   7.5573  19.7267
#>    3.2061   2.1481   3.0787   0.8387  -7.9401   3.8010  10.1523  -8.9512
#>    4.8704  -1.4381  -2.3305  -8.0051  -6.7356   0.0588  -2.4345   5.2877
#>   -1.1401  -7.1325   7.7245 -11.2140   3.9308 -11.3877  -6.8392  -8.5941
#>   -1.7634  -6.0794  -1.7959  11.0061 -17.3828   2.6756   3.3136   3.0608
#>   -1.9718   5.6633   2.5086  -0.9622 -10.8186  12.7519 -14.9045  -5.0209
#>    0.4394   3.5905   8.4949  11.6114  -0.0725  11.8052  -0.3338   0.7985
#>   -5.6926   0.4183 -21.4916   5.5453 -15.1921   1.5979   6.3087  -5.6916
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]