Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -7.3448e-01 -1.2362e+00 -1.4581e+00  1.2142e+01  1.3595e+01 -2.4496e+00
#>   4.8957e-02 -3.3140e+00 -6.8805e+00  8.5988e+00  2.3037e+00 -1.2164e+00
#>   1.5045e+00 -8.8831e+00 -1.6469e+01 -5.3374e-01  4.9188e+00  5.3270e+00
#>  -9.1954e-01 -2.7045e+00  6.5634e+00  7.9245e+00 -1.0957e+01  2.0385e+01
#>  -3.2893e+00 -1.8741e+00 -1.9460e+00 -1.9213e+00  1.7968e+01 -8.3825e+00
#>  -4.9716e+00 -4.0186e-01 -5.0536e+00 -5.4406e+00 -8.7079e+00  5.4079e+00
#>  -2.3525e+00  8.3421e-01 -2.4756e+00 -1.2536e+01  2.7002e+00 -3.8491e-02
#>  -3.4170e+00  2.1147e+00 -1.4948e+01 -1.8947e+00 -8.5367e+00 -1.2640e+01
#>  -6.2221e+00 -6.1952e+00 -1.4751e+01 -2.3797e+01 -6.1038e+00  5.7564e+00
#>  -2.1276e+00  7.8427e+00  1.6603e+01  5.9204e+00  5.7183e+00 -9.3560e+00
#>   8.3240e-01  7.7667e-02  1.2023e+01  5.4727e+00  5.1958e+00 -4.3252e+00
#>  -2.5298e+00  2.6866e+00 -6.3151e+00 -1.1782e+01 -1.6691e+01  6.8662e+00
#>  -5.6060e-01  3.4799e+00  1.1357e+01 -4.8728e+00 -1.0041e+01  1.1226e+01
#>  -2.8960e+00 -3.9595e+00 -6.3023e-01 -3.4463e+00 -1.1789e+01 -2.2860e+01
#>  -2.2335e+00  7.2580e-01  5.8281e-01  1.7076e-01  7.1918e+00  4.9005e+00
#>   1.8543e-01  1.4993e+01 -3.8425e+00  4.2354e+00  6.5017e+00 -3.8590e+00
#>   1.2040e+00  7.3195e+00  6.6233e+00 -2.8392e+00  1.3865e+01 -3.3651e+00
#>   4.2587e+00 -1.0938e+01  1.2238e+01  1.9905e+00  1.3579e+01 -1.0856e+01
#>   5.2014e+00 -1.8562e+00 -4.1503e+00  4.3198e+00  6.9398e+00  9.7186e+00
#>   1.8542e+00  5.9770e+00  1.0752e+01  9.8820e+00 -6.7867e+00 -4.2807e+00
#>  -6.7201e+00 -6.5871e-01  5.7063e-01 -2.6551e+00 -2.9596e+00 -3.4035e+00
#>  -3.5361e+00 -1.1075e+01 -1.1347e-01 -1.1741e+01  4.2205e-01 -6.0658e+00
#>   6.1692e+00  1.0535e+01  3.4509e+00 -6.4695e-01  7.1486e-01  4.5980e+00
#>  -6.9357e+00  5.5241e+00  1.6477e+00  3.6907e+00  1.0945e+01  1.8864e+00
#>   6.8203e-01 -9.0886e+00  2.0783e+01  8.8343e+00 -1.4208e+01  8.7365e+00
#>  -4.8361e+00 -2.1113e+00  6.0010e+00  1.8338e-01  4.6250e-01 -1.0242e+01
#>   9.4560e+00 -4.0450e+00 -2.6311e+00  9.2974e+00  8.9021e+00 -1.9001e+00
#>  -1.1599e+00  8.4804e-01  9.7674e+00  5.0093e+00  3.8889e+00 -8.6818e+00
#>   1.0776e+00  3.4562e+00  9.9498e+00 -9.6968e+00 -2.2631e+01  1.9924e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]