Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.5448   1.2901   1.5434   6.3455   1.4009  14.8432   1.6724  -7.4952
#>   -2.5230   5.3240  -7.1234  -0.9344  12.7678   6.9636   9.1123   2.0262
#>   -2.3053  -7.3659  -1.3950  -3.2961   7.9112   2.1792   0.1904  18.1738
#>   -2.1117   6.6698   5.3199   1.9880   2.4330  -2.8386  -7.2043  -6.5615
#>    1.2127  -2.7983   3.3090  21.4839 -11.8679   6.0753  -9.2123  -8.5521
#>   -2.6392  -2.0917   2.9758  10.0814  -2.0469  17.4157   3.9683  -1.8868
#>    1.9913  -0.1197  -0.5615   9.1516   7.9959   1.7823   4.4261  -0.4376
#>    0.2732   1.1555  -9.1663 -11.5864  -3.6409   8.4345  -3.3481  19.1384
#>    0.7650  -7.2553   6.7952  -2.0084   2.1724 -13.8476   8.9356   0.5706
#>   -2.4313   0.9998   3.9196  -4.5359  -5.2435   2.1436   5.8370  -2.7249
#>   -7.7498   3.6764   1.2874  -4.7221  -1.2992  -1.3159  -6.7450  12.2094
#>    2.4754   0.8525 -10.3894   1.9759  -8.0504   6.0570  -1.5329  -6.5467
#>    1.0640  -2.8211  13.6959   5.9376  11.7679   4.4604  17.4675  -6.6158
#>   -2.2735  -8.1078  -3.1315   3.2537   5.9160   7.1138  -1.9528  -3.0274
#>   -0.6155   0.8290  -4.7409  -5.6425  -5.5552   6.2029   0.5173  -7.1312
#>    1.6752   4.8781  -2.3429 -12.3772  -0.4987  -5.8340   3.1777   7.8707
#>    3.7060   1.2426  -2.7286   7.5446  -7.7543  -8.6475  -6.5498  -6.6602
#>   -3.4022  -5.2043   2.0431  -9.0702  10.0579  -3.9457   4.1923  20.0478
#>   -4.3403   0.5250  -9.0561  -9.2854  17.7088  15.9237  10.4263  12.2533
#>    2.8152  -9.6371   3.2834 -10.4323  -6.1575   3.8230   7.1056  -6.2471
#>   -2.4253   0.2299   1.8568  -5.0530  12.8347   8.2696   7.2793  13.5315
#>   -4.0278  -1.9836  14.7083   7.8173   1.5483  -2.6561  -7.5900  -7.9916
#>    0.0173  -1.3612   1.0978  -2.3783 -11.2442  -0.9578  -0.0555   5.8607
#>   -2.2095  -4.9430  -2.5923   5.8974  -0.4771  -7.4573   1.3170  -3.8364
#>   -4.9371   6.9134  -5.1633   0.3576  -1.3693 -13.8869   0.8309  -4.6487
#>    1.4200 -12.7995  -8.6894   7.8059   6.1141   4.8828   7.3907   8.0135
#>   -0.9922   7.1333  -1.8521  -8.0107   8.6744  10.7032  -7.4985  -8.5607
#>   -0.0819   1.6974   0.7232  -2.9522   2.0768  11.5902  -9.1550   3.6657
#>    3.6314  -2.4210  -1.5347   0.7991 -23.2950   5.9835  -0.8582   1.5833
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]