Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.5664  -0.5645  -5.5421  -2.3602  15.3819  -6.8417   7.7505  15.7845
#>    8.5446   1.8762  -6.3470  -5.8164   0.6391   8.4583   0.5222 -13.7666
#>    1.6134   4.7398  -2.4490   2.3162  15.6180   4.2514 -11.7163   1.9135
#>   -8.8422  -7.5241  -2.5724  -0.9804  11.8932  17.6931   7.2638  -0.5295
#>   -1.0091   8.6127   0.0639 -10.1194  -7.7069  -1.7400  -7.9723  -6.8718
#>   -9.4260  -7.1009   4.0495  -1.8468  11.3965   3.8607   9.2896   8.3964
#>   -4.4624   6.1650   8.6440   1.0479 -11.7659 -10.7817  19.9293  -4.7242
#>    1.7991   0.3722   2.8908  10.2188  10.7319  10.0572   4.5061  -0.0561
#>   -3.5874   5.3135  -1.5458   0.5571   2.5330   4.9231   2.8055   7.2210
#>   -0.4421  -2.6635   7.6063   8.9517  21.9115   5.8036   5.3622 -16.9509
#>   -3.5201  -4.5030  -8.7695  -3.5425   4.2825   4.0456  -6.6372  -2.3910
#>   -3.0683   3.0093  -0.3674   9.3055  11.6837  14.6645   1.5809  -0.7298
#>   -6.6300   1.4148   3.5000  -7.2363  -2.7412  15.4995  -0.8540  -2.7832
#>   -5.7575  -7.8188   0.1476   0.2437   3.6774   8.9728   4.8823   6.3646
#>    3.6570   4.5374   2.9662   1.1164   1.8398   1.6675  -2.6432  14.6931
#>    2.2534  -4.6350   6.5835   2.3575   0.7011   0.1763  -2.6895   0.9023
#>   -1.9076  -1.0813  -0.6385  -3.1278   2.2959  -1.9939 -13.8303  -3.7038
#>    4.7401  -0.9031  -3.1256   3.3669  -4.6462  -7.3639  -6.1438   3.9139
#>   -1.6732  -4.9505  -9.3800 -14.9602   0.7270  -0.6454 -16.7655   2.2615
#>   -1.7154  -1.4044   6.6069  -9.4290 -20.9012  -3.5358  -3.7543   1.4444
#>    3.7557   3.5249 -11.3824   7.5408  17.4743   9.3881   6.5255  14.2724
#>   -1.3434  -1.7034  -0.0067   2.5786   4.5176  -4.9764   3.7948  -0.3935
#>   -4.9994  -3.9440  -0.7089  -5.1773 -11.0129  -1.1877  -7.4364  -0.9528
#>   -2.0891  -5.0754   1.7436  15.8696   8.5253 -11.1032   8.4862  -3.6599
#>    2.6928   4.9310   2.8103  11.3924 -13.0817   8.1577 -10.2293  -3.2307
#>   -2.0934  -0.5840   0.5365 -16.4291  -5.9302 -16.7438   6.5463  -6.3323
#>    2.5633  -0.0798  -2.7940   5.9361   4.6370  -3.5298   8.9842  15.9526
#>    6.4969   6.8524   0.0132   1.4622  10.4987  -3.9717   3.6281  -2.9226
#>   -1.2996  -1.0767  -4.1298   5.6311   8.1495  -8.6921 -20.0916   8.7176
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]