Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.7013   0.6125   5.7400  -4.2543 -19.9561  10.9571  -1.8786  18.1239
#>    2.5765  -4.2738  -4.1109  -3.0700   4.3029  10.9830   6.8912   5.4388
#>    1.5126  -4.1140 -16.3651   8.8167   3.1406   8.6806   0.6065 -15.4951
#>   -1.6061   5.3407   3.3924  -8.4205   6.1786  -9.2309  -7.2697  11.0396
#>   -0.1792  -2.3599 -12.5488  -0.5387  -9.7220  -9.2788   0.0953  -2.0306
#>    2.5224  -3.3488  15.2930   8.5366  -8.5472 -22.8318 -24.4504  -2.9244
#>    2.4025 -10.4316   5.0918   9.2661   2.8645   6.2502  10.2489  -2.8719
#>    1.7329   3.4411   7.5458  -9.1820  -9.0016  10.4400   0.2861  -5.5539
#>   -0.2649   6.1542   4.0191   5.1298  -5.0282   5.1138  -2.0441  -0.8243
#>    5.0738  -3.3002  -2.4836  10.6075  -1.3035  -6.8131  -1.9149  -4.8800
#>   -3.5340  -1.4363  -7.3337   4.3148   7.8537  -1.1151   0.5349   2.5072
#>    3.0516  -6.0011  -3.5568  -9.4548  -8.7207   6.7580   3.1695  16.3992
#>   -5.5107   3.1483  -1.6168  -3.0018  -7.3854  -0.5795  -8.4748   0.6282
#>    0.8497  -3.7677   5.4871   2.7847  -5.3826  -2.1525 -20.8961 -17.4950
#>    0.8096   3.0226   2.4084  -1.8942  -4.1514  -5.3720  -6.1207  -2.6430
#>   -7.1683   6.5739 -12.0245 -13.2622  -8.5366   6.1983  16.0593   3.8246
#>    2.8515  -7.3733  -5.6940  -6.0153   7.1944   6.7437  -8.2807   0.1110
#>   -2.7445   0.9877  -5.8100  -4.5834   2.0650  10.8380  21.7354  -3.7583
#>   -0.4665  -7.0780  -8.4939  -8.8193  -2.3350   3.5871   3.8817  -7.4528
#>    2.1305   3.9615   1.6674  -4.6999   4.8171  -9.3951 -18.3553   5.4740
#>    3.1440  -5.2698  -3.8399   5.3436   1.9591  -1.0932   1.6786  -0.2328
#>    5.5267  -4.9073  -0.4597   7.1531   3.8799  -0.2166  15.1827  -5.3005
#>   -0.5368  -3.8162   3.6793  -6.3057   2.7714  -3.3201  -6.1980  -8.3054
#>   -0.9413   2.2534 -10.5605  -8.6020   4.7152   5.1267  14.7347   6.9569
#>   -4.5219   2.4171  -9.6658 -16.4020  -5.4818   0.7714   4.2823  -0.4565
#>    1.9162   2.2344   3.0384 -14.7793  18.2542   9.7453  11.2981  -0.4316
#>    0.4576  -0.5850 -11.4139  -3.2408   2.1112  -2.5823   5.0971   2.5660
#>   -0.7295   8.5391   3.9659  -0.5196  -7.7032  -8.5856  10.4621  -4.5074
#>   -2.2997  -0.4591 -11.9689  -6.9738   1.7914  15.7775   9.0226  -1.4905
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]