Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.1919   4.1692 -28.8431   4.0672  -1.9838  -0.9573 -11.1987   7.3520
#>   -0.4289   3.8766 -11.1825  -2.4893  -9.0904  -8.1631   8.5162   5.5628
#>   -3.1465   4.5818   2.5105  -1.5803  -8.4826   5.3410   3.5953   6.3041
#>    0.7585  -6.5036  -1.1462   0.8035  -7.4552  20.5874   4.6943  -7.1702
#>    3.4390   6.2673 -12.3106   6.9186  -2.9557   0.2433   0.2098  14.6131
#>   -7.0289 -10.3977  -3.8581   2.2643  -0.7350  -2.9434  12.9950  -4.6229
#>    1.9060  -2.4508 -11.8614   9.6193   5.4064  -9.4666  -3.5253  -2.5050
#>   -1.6861   5.0584   1.3806   1.5576   3.8945  -0.2434   1.8209  15.4318
#>    6.3441  -2.8367  -5.6430  13.4444 -17.5929 -15.5508  -4.9466 -15.9238
#>    1.8186  -0.4719  -9.0380  11.8499 -14.5617   2.3109  -4.9976  -5.1086
#>    4.7827  -7.3399   1.9310  14.1898   2.4104   6.3088  15.4068   0.3991
#>   -2.0691   6.3117  -0.0188  -1.6421 -14.0239  -2.7686   0.5563  12.4278
#>    6.1596  -7.3588  -1.3298  -4.4039 -14.4924  -9.5417  -6.5300   2.3134
#>   -2.7265   3.7108  -0.6443  -5.4496 -19.4264  -0.1787  -4.4145  -4.7817
#>   -0.7560  -3.0001  10.0789  14.7553  -7.3185   0.0939   5.4801  -5.6896
#>   -2.4907  -2.0216   4.2815  10.2909  -2.3423  -5.6793   3.6370   6.4363
#>   -0.0687   1.5759   6.8357  -1.5023  11.5234   3.7965   7.3749  -6.0008
#>    1.4119 -10.3798   4.1580   0.0340  -3.8740  -0.1238   3.2789   3.5722
#>    0.6976   8.7419   2.0718  -0.9456 -10.4216   7.1485   2.2369  -4.7817
#>   -1.1489  -1.2928  -0.0313   1.9704 -13.6334  -6.8503  -5.2433   8.3019
#>    0.5082   6.5937  -1.9546  -1.7839  -2.1009  -5.0943   2.8574   3.5631
#>   -3.6515   4.0975   0.8678  -0.4665   5.3781   1.7046  -8.8076   3.9119
#>   -0.9477   8.2014  -3.7118  -8.7298  10.4113  -1.7346 -10.1700   0.0022
#>   -2.7802  -8.2172   0.2834  15.7253  -6.2562  -0.8723  -3.6359   5.1345
#>    3.7375   3.0126 -13.8703   0.4506  -6.8408  -8.5333   9.0485  13.2216
#>   -2.5036  -6.2540  -6.4757   2.3914   1.6869   3.7813  -8.0514   0.2586
#>    2.1867   5.9309   8.5212  -6.3513   1.4131  -4.7476   4.0878  14.3700
#>   -1.5434  -0.0607 -14.4874   1.2763  14.3376   2.1017   8.9879  -4.8597
#>    0.8047  -2.2099  -8.1483  12.0790  -9.6308   4.5172  -1.4728   2.0988
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]