Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  1.9197e+00 -1.2312e+01  3.0959e+00  2.7604e+00  8.4807e-01 -3.4773e+00
#>   1.0078e+01  8.6088e+00 -5.4258e+00 -1.2265e+01  2.4311e+00  1.2589e+01
#>  -9.2067e+00 -1.2402e+00  1.9466e-01 -1.0464e+01 -1.5369e+00 -5.1264e+00
#>  -6.7908e+00 -7.8579e+00  2.5693e+00 -4.6473e+00  5.4138e-02 -1.1053e+01
#>  -1.5469e+00 -7.0676e+00 -9.8083e+00 -1.6510e+01  6.8042e-01 -9.8394e+00
#>  -7.5923e-01  4.6964e+00 -9.0567e+00  1.3747e+01  3.7052e+00  3.8230e+00
#>   5.6615e-01 -2.1482e+00  1.4787e+01  1.4492e+01  1.1638e+01  2.0596e+00
#>   4.3421e+00 -6.5926e+00 -1.7486e+01  1.2183e+01  6.6251e+00  1.1502e+01
#>  -9.7654e-01  7.8203e+00  8.4649e+00  1.5065e+01 -3.2814e+00  4.9624e+00
#>   1.9271e+00 -7.3882e+00  9.9982e+00  7.9823e+00 -6.3816e+00 -4.1590e+00
#>   2.1923e+00  1.7262e+00 -4.7152e+00 -7.3823e+00  1.7271e+00 -1.3161e+01
#>  -3.7278e+00 -4.6311e+00 -1.3024e+00 -3.2499e+00  2.1188e+01 -1.3477e+00
#>   1.8437e+00 -8.3230e+00 -5.4992e+00 -1.0539e+00  1.1240e+01  6.2070e+00
#>  -6.5238e+00 -1.3901e+00  3.4133e+00  4.1721e+00 -4.8087e+00  4.4298e+00
#>   2.7507e+00 -2.5523e+00  9.7905e+00  2.5079e+00  2.8052e+00  2.3202e+01
#>   3.5336e+00  1.0482e+01 -1.3317e+01  6.3614e+00  2.0875e+01 -1.0982e+01
#>   1.7576e+00 -3.5877e+00 -7.3325e+00 -1.6784e+01  3.7258e+00 -1.2854e+01
#>  -2.5838e+00 -4.6098e+00  2.9866e+00  3.8808e+00 -7.5763e+00  1.1641e+01
#>   1.4743e+00  5.7789e+00 -1.4953e+00 -2.1835e-01  1.7466e+00  7.9104e+00
#>  -4.9223e-01 -3.5151e+00  3.2671e+00 -7.3098e+00 -5.6433e+00  9.0557e+00
#>   5.9600e+00  6.7135e+00  6.6036e+00 -6.4395e+00  4.1050e+00  4.0772e+00
#>  -5.3949e+00  5.5183e+00  2.6613e+01 -1.8485e-01  1.3179e+01 -1.7889e+00
#>   1.4424e+00  1.3779e+00 -5.2606e+00  1.6759e+01 -3.5459e+00  8.5162e+00
#>  -2.9207e-01  2.3793e+00 -6.5067e+00 -8.5767e+00 -6.8852e+00  3.1619e+00
#>  -1.9904e+00 -6.5453e+00  7.2891e-01  2.0893e+00  1.0109e+01  8.2175e-01
#>   6.4377e+00 -5.2445e+00  7.7791e-02  1.7538e+01 -6.1960e+00  2.5948e+00
#>   1.3253e+01 -4.1163e-03  9.0494e+00  4.7512e+00 -8.5202e-01  2.2299e+00
#>  -2.9660e+00 -4.6916e+00 -1.7682e+01 -2.0462e+01  5.4691e+00 -8.1678e-01
#>   1.7755e+00 -2.3381e+01 -1.4763e+01 -8.7063e+00  2.7812e+00 -1.6440e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]