Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.6646  -1.8094   5.2286  -3.4416  19.5120  19.2841  -5.8100  -5.7083
#>    2.8254  -6.2662  -3.1094  -2.5426  -6.3099 -15.8192   1.8576  -3.8112
#>   -2.2294  -1.5311   5.8453  -2.5372   8.2897  -5.2018  -5.7763  -4.9294
#>    2.7433 -10.0825   5.6772  15.0867  -6.3636  -7.8383  -1.3403  -6.5818
#>    2.1539  -4.2675  -1.1012   5.0492  11.5843  -4.3779  -4.9258  -4.4109
#>    1.1161  -5.7472  -2.6541  -9.0446   8.1375  -1.1710  -4.6689   3.9913
#>   -1.5926  -5.7593 -11.8773  -3.6910  -8.2089  -7.2175   0.6352  10.5451
#>   -1.6047  -2.5445   5.1382   3.1560   8.0551  -1.0450   4.8106   4.8787
#>   -0.6889  -4.8968  -6.0302   4.8386  -8.4264  15.2081   1.9552   6.8326
#>   -0.0782  -1.9864   0.4439  11.5343  -9.1718   3.4562   3.0318 -11.9128
#>    0.5197   0.0842  11.6424  -0.4578  -0.0415   9.1981   8.3534  -6.1817
#>    5.1058  -5.6014   0.3660   2.2947  23.4400  11.1427   9.1648   5.1085
#>    0.4373  -9.8314   2.4296  -6.2814   1.6733   1.0334  -1.1881   3.7591
#>   -1.3131  -1.3331  -1.6499   3.8448  -4.3623  -0.5050  15.4422  -2.6879
#>    1.7166  -2.3380  -4.3102  17.9628   0.4073  -1.8867  -8.2553  11.8142
#>    0.0657  -5.0380   3.3284   7.0214  -3.4765 -12.1388 -15.4199   9.5583
#>    0.6198 -11.3078 -13.4936  -4.1233  14.9186  -1.7795   1.3303 -13.2161
#>    3.7725  -3.0204   3.1244   5.6665   7.4473  -2.6822   8.4349  -4.8011
#>    2.5141   7.4952  -8.2470   2.9325  -7.7879 -12.8596 -22.0333  -4.6027
#>   -2.3652  13.2076  -6.3608   5.1585  12.6463   2.9881   1.4385  -2.2884
#>   -0.2563   2.1453   0.2159  -2.9039 -13.6883  -9.0454   6.0830   5.0878
#>    1.3672  -5.7134   8.0859  -1.8634  -7.5642  -7.6660  -1.6010 -15.7323
#>   -0.8277  -1.6031 -15.1630   3.1740  -4.6587  -4.8044 -10.5754  -2.2191
#>   -0.5649   3.8140   5.4015  -4.1949   1.6703   4.9231  11.3898 -16.8064
#>   -3.1534 -10.1632  -4.2029  -9.1569 -13.6275   6.1201 -14.6758  -6.1401
#>    3.0250  -4.8907   0.5886  -4.1673 -12.2965  -2.1392   3.7195  -4.1866
#>    3.0092 -11.0607   4.9042  14.7157   0.7340  -6.4815  -9.8901   7.6259
#>    1.7981  10.0526 -11.6081   3.8351  -5.0862   0.4934  -9.7423   1.9444
#>    4.0456  -1.6318  -2.4735  -4.9799   0.4247 -15.2109   2.0268   1.4802
#> ... [the output was truncated (use n=-1 to disable)]
#>   -1.0408  -3.7990  14.3224   5.7342  -2.9311  -2.6270