Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.6921   9.0173   4.9147  12.1982  -0.1162  -3.0589   8.7984   0.1404
#>    0.5076  -1.7511  12.2738  10.0242   1.9376   5.5108   7.8075  -7.1888
#>    1.0884   9.8051  -3.9313 -10.8721  14.3435  -6.9888   9.8922  -5.0453
#>   -0.8968 -11.8598   0.5657  -3.3580  -0.9035  -8.7698  10.0191 -10.8418
#>   -1.1493  -4.2375  -4.0060   3.4437   3.9416   8.0358  -8.5217   2.0901
#>   -0.4300  -5.1054  -0.2288  -4.1018 -12.9633 -21.6224  -2.7375  -7.0377
#>    1.6348  -3.2123   7.5097 -11.1094  -3.4373  -8.4069   5.1785  -2.2025
#>   -0.6977  -0.9418  -4.1660  -7.1729  12.7961   8.1835   0.1557   2.2718
#>   -6.4029 -10.3086 -10.6119  -4.5414  -6.1056   0.5732   5.0110  -2.7040
#>   -2.2762   7.6906 -14.4088  -7.0084 -10.0523  11.8428   8.5148   4.4882
#>   -3.1116 -11.4370  13.6079 -14.0305  15.5007 -11.3986  -5.3768   3.4804
#>    3.7441  -1.5624  -4.3871 -18.1451 -15.3139  -5.7717  11.9691   7.9437
#>   -3.8526   2.1778   4.5112 -11.8837   9.5710  -1.3150   2.9580   2.7567
#>    3.5297   3.5429  -8.9990  -9.0771   7.7655 -15.3564   6.0650  12.3424
#>   -0.6278  -4.8145 -11.1419   2.7091   2.5462   9.4055   4.0454   3.2583
#>    0.8023   3.9420   0.4193  -1.9721  -0.0431  -7.0690   3.4788  -2.5186
#>   -0.5506   7.2752  10.8379  -2.7186  -9.2422  -0.7835  -1.3538  -9.9522
#>   -3.1442 -13.5001   2.3077  -3.6462  16.2478 -10.5980   2.6128  -5.8626
#>    1.6928   3.4782  13.7792  11.1303  10.5211   3.1453   5.3208 -10.7673
#>   -3.2290  -2.7818  -6.4893 -16.4697   2.9140  -4.9530 -15.6964  -6.9806
#>    4.0218   4.7076  -7.3530   0.1485  -1.3094  -1.2186  -5.2101  -7.9262
#>   -4.1213   1.1775  -0.0613 -10.7009  11.0970  -0.3248   9.2904   4.4042
#>    1.6164   2.0790 -10.2825   0.3957   1.2459 -13.4779  -4.6139   8.3315
#>   -5.4698  -5.4139   8.9969   0.3097  -0.8334 -11.3563  -0.2946  -9.5958
#>   -2.0613   0.5447  -8.5085   8.7870  -9.3668  -2.2592  -5.2660   5.6451
#>    1.1770  -1.3697  -0.4844  -2.4537  11.6103   5.1244  -0.7211   6.2045
#>    1.4614  -1.7045  -8.0803   3.0367  -3.6801 -15.3633  -5.1528  -0.2938
#>    1.2239  -0.2401  17.5869   5.2295  -7.4983  -3.9174  12.5799  -5.6311
#>   -0.5164  -4.9092  13.7974   8.8383  -5.0418  14.1127  -2.9130  -6.0256
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]