Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.9809   9.4154   0.2932   5.0198   0.7690   3.4691  -8.3080  -4.6249
#>    1.4817  -3.3067  -3.7462  -5.4399  -1.3110  -7.1353  -9.9053  -0.1885
#>   -2.9807   0.2900   3.9238  15.0058  -6.5330  12.2541  -3.8824 -25.7126
#>    6.2793   8.6631  -0.6012  -3.8531   1.0107   7.2639  -0.2888  19.0648
#>    1.5937  -2.2887   4.5719  -4.0389  -5.2190  -5.6768   9.6091   3.7120
#>    0.7131  -2.9165   3.9254   2.5088 -20.9039  -2.6128  15.2187   4.5515
#>    2.2357  -7.9762  -1.2694  -3.4651   9.0179  -0.2224   2.6533 -16.3798
#>   -3.9930  -0.8668   3.1539   6.0439 -22.7263  -5.7074   5.1207   1.2662
#>    1.7931  -6.7163  -1.5861   1.7877  23.5932   7.4528  -5.6498   8.6956
#>    5.8168   2.3163   1.6376  -2.5620  -1.6998   2.0117  -2.1798   8.3321
#>   -3.0878   5.2433  -4.2040  -6.0574  -9.5782 -12.4961   5.7918  -1.1286
#>    2.1138  -1.7401  -3.2358   9.6442   6.1512 -12.1194   2.9391 -22.0491
#>    6.4040  -3.7813  -9.2249 -23.0060   6.8803  -3.1282  14.7892  -1.6255
#>   -1.5059  -1.3862  14.8461   8.4618   2.0534  -2.9406  -7.0480  12.7619
#>   -8.4038  -3.7526 -15.5828 -10.9367   9.5051   4.1633  20.8943 -14.3828
#>    1.6204   2.6895 -15.2136 -20.5659  -0.5727  27.2431  16.9685   7.8624
#>   -2.8687 -16.5252  -1.4384   9.1138  12.3715 -17.5314   3.2414 -14.5445
#>   -4.6157   0.9405   0.3637   9.0956  18.4015   8.0061   7.2220   3.9499
#>   -1.4245   8.1250   8.1558  -9.9134 -19.6816  -6.9693  14.3736  13.6753
#>   -4.6720  -0.5144  -1.0481  11.0620  -0.1061  -4.3864  -9.9134  10.6606
#>   -0.9771  -1.9525  -0.7638   3.9144  14.1234  -0.5184   4.1215   4.6306
#>    5.0566   5.2119  -3.6151  -6.6945 -14.4937  -9.6666   8.9326  21.0732
#>    0.6914   0.3957   7.2853 -14.1084   1.0298  -5.2033  12.2023  10.1409
#>    4.0829 -12.8849   9.4885  -5.3073   9.7170  -0.3481   2.1793  -9.6620
#>   -3.0163  -2.4871   8.4913  -8.3809  -1.6861   0.7691   1.3524  -5.1497
#>   -8.1727   2.3474  -2.6086   4.6774  -6.6629   0.5830 -22.7847   4.8963
#>    1.2838   0.2801   2.9005   0.5925 -15.6888   5.1501   6.0367  -0.6404
#>   -8.7404   4.8849   1.0832   6.5438  -9.6434   4.2975   5.7335  -9.2730
#>   -2.8992  -1.4414  -3.0572   5.8609  25.4146 -15.9775  -1.3906  -2.8351
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]