Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.9962  -6.8410  -9.7206   2.0912  -5.7290 -13.8295  16.5079   3.6231
#>    2.2828  -6.4220 -12.6670   0.3073   0.9716  -0.4576  -2.9017  13.2712
#>   -4.3868  -3.3470  -5.6179   3.8587  14.2506 -14.0854   1.3766   9.6349
#>    1.9361   2.3375  -2.1790  -3.3186  10.1626  -0.2841 -12.3840   6.2410
#>    1.2206   2.1269  -0.2283   7.5235   5.9889  12.3955  -4.3640  -8.4516
#>    0.2132  -3.9637   3.6489  -8.5017 -18.5674   3.4276 -12.8156  -2.7024
#>   -0.0783  -0.3071 -10.1867  -3.8818  -7.4861   9.3433  17.1803   6.0149
#>    7.1808   9.8899  -2.8179  -5.6840   4.7684   1.6562 -15.8096  10.8847
#>    0.3926  -7.3797   9.0103  -6.6619   2.1771  -9.4027   5.8511   2.4427
#>    3.1119   2.6502  -0.4337   0.2627   9.9691  -2.3813  -5.6400  -2.5955
#>   -3.7383   0.4160  -6.5404  -3.5911  -5.8637   1.8159  -6.0698   8.2499
#>    0.0246  13.0606  -4.9455 -13.9440  -2.6106  12.8329   4.9852   6.8887
#>   -0.7408  -5.2205   2.0930  -0.6304  -4.8601  -1.3401  13.8997 -10.2049
#>   -1.6386  -6.4757   5.3172  19.5993  -9.6224   7.3275   3.4691  20.0180
#>    1.9013   7.0793  -1.8914   0.5193   4.0908 -14.9432  -9.9961   4.3125
#>    0.5859   8.4451  -5.3286   2.4643   2.8540   6.2242  -1.1521  -3.3420
#>   -4.6864 -10.6282  -5.1643  -7.5451  -8.8455  11.3403   5.4814  -8.1942
#>    4.5380   0.5140   7.3628  10.6066  -2.0442 -11.3638  -1.9618   4.0357
#>    0.7788  -4.6237   3.2347   2.0156   2.4450  -1.9436  17.3203 -11.4665
#>    3.8521  -1.2819   6.1252  -6.1926  -3.4093  -2.3444   2.8324   9.7142
#>  -11.7901  -1.3050   0.0958  13.6131 -13.5288  -4.5046  10.8927   4.5111
#>    6.8397   0.7730   2.4411  -4.5399  -6.6518  10.9783  -8.8917  -2.5796
#>   -1.6880  -2.4569   6.8443   9.1627  -3.3440 -11.9651  -1.4052  11.0695
#>   -2.8530   3.6376  -5.8881   5.7604  -7.2003  11.3840 -13.3403  -5.7271
#>    5.7373  -3.4428  -6.5271   3.1254   0.6100   8.1997 -13.0621 -10.7866
#>    2.9637  -2.4079 -10.4895   9.1595   2.9783   4.1790 -13.1832  -7.9807
#>   -5.0767  -7.6248  -8.1771   9.7599   6.0867   9.0273  -1.7308   9.5950
#>    4.3108  -1.8128  15.9549  -7.1023  -4.8195   7.9207  -2.3597  -3.4277
#>    4.6056   7.3382  13.1486  -9.8649  -8.5041   8.7796   7.1205  -3.7983
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]