Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.9249   2.4293   4.5660   2.6915   5.9638   8.6202  18.4712  15.0249
#>    1.6916  -4.2858 -10.1143  -8.5957  -7.2480 -14.1939 -13.5813  -8.0127
#>   -0.0501  -4.2798  -7.2471  16.2155  -3.2242  -3.3155 -10.1979  16.8578
#>   -3.1282   3.7004  -0.1223  -5.3052  -7.2743  -5.6941 -11.0171 -13.7992
#>    2.0608   3.3909  -0.4471  -9.1662 -11.2572  -0.7975  -9.1122 -10.1435
#>    1.8110  -3.3310 -11.1561  -6.3903   3.6697   6.1865  -1.5491  10.9944
#>    0.1500   4.4568   7.5358  -8.2705   8.6120   4.3711  14.2621 -17.2545
#>   -4.3548  -5.1931  -1.5842  -7.0143  -5.3047  -4.0054  15.3604  -3.8682
#>   -2.4988  -3.8740  -2.0964   6.0484   3.8575   0.1416  -3.2313   2.1438
#>   -5.0963   0.6466   3.2928  17.0933  -8.6719 -13.2534 -11.3316  12.4109
#>   -0.1515  -0.6691  -4.9399  11.7885   2.1078  -6.1335   6.4144   5.8485
#>    2.7124  -3.0057   4.0538   2.3739  18.1565  -6.5129   7.7312   6.9501
#>   -7.8388   0.8437  -5.9675 -14.9688  -2.7702   9.3217 -21.1293  -2.7878
#>    4.0894   2.3286   5.0760  -0.9346  16.4044   3.1466  -9.9923   9.6929
#>   -3.7741  -3.1453   5.0875 -11.6707  -7.9851   1.6691 -17.6684 -19.6629
#>    3.2929  -5.2655   0.7860   3.9041   0.8772   0.4802  -8.6495   0.4751
#>    0.5398  -1.0929  -2.4384   3.4308  -8.0340  -7.5538  -7.8538  -7.7657
#>    4.6763   8.4068  -4.3099  12.0293   4.6455  15.0426  12.0842  -5.7419
#>   -0.2327  -2.0054   1.2107  -6.9552  -5.9942   0.4654 -16.1192   0.9307
#>   -0.4243   1.2434   2.8767  -5.8644  -1.4287  -6.8593   5.7689 -14.6095
#>   -5.8347  -4.3654  -3.0589  12.0194   8.6153  -7.3350  -2.9851  15.3872
#>   -1.2045   6.3810   1.3005 -21.6877   4.9287   4.2613   4.6565  -0.9844
#>   -0.2091 -10.7118   1.6514  10.6687 -11.4588  -1.3772  -2.6919  -4.1890
#>    6.1771  -0.9034   5.1456  -4.2521  -1.9989   4.7975  -6.3208   0.7503
#>    0.4927   5.9526   4.9048  -4.3457  -4.6616   3.1915  30.2209   4.6729
#>   -0.8637   3.0421  -1.2320  -2.7820  -4.4007   5.4228  -3.3856  -8.0031
#>   -2.9544   5.0471   3.1934   5.7427   0.9489  -5.2870   7.7020   5.5854
#>    4.0382  -6.3310   2.0184   0.9051   1.1474 -12.0427   5.0655   7.6731
#>   -4.2078  -2.6056  -2.5428  -4.4402   5.4927 -12.1059  -1.2314  -2.4047
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]