Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.3379   3.5725  -5.7148   8.1925   6.4710  -4.0041  -2.8132  -7.2028
#>   -0.4870  -4.2542   4.7555   6.0299  17.4717  25.9075   3.2213   6.6614
#>    6.2246   6.3465   5.9459   8.3223 -17.4084   0.4110  -9.7989   1.7102
#>    1.8347   1.4888   2.7489  -3.8783  -2.9766  -6.2077  -6.8106  -7.5902
#>   -4.9048   2.9454  -5.4741  -3.7689  -2.7052  -5.1289  -9.8708  -5.2888
#>   -1.7240   4.8995 -10.0904  -5.4459  14.2236   0.6262  -3.3011  10.4533
#>    1.0256  -3.0095   3.9680  -3.1247   0.9452  -2.3982  -4.6501  -2.9356
#>   -1.7842  -8.1044   7.7617 -15.6447  -4.1471  -7.8902   7.6600   2.5758
#>    0.1193   2.4650   0.8524  -0.5137  26.9243  -9.8606   6.4235  15.0213
#>    1.2682  -4.8371   4.2884  -1.8733  -0.7039  16.7134   5.6588  -4.3423
#>   -6.2096   3.7542  -1.9195   1.7884   6.7693   8.1212   4.5935   5.1584
#>   -3.1994   0.2680  -6.1408   9.7095  -0.4451  -0.2235  -4.2832 -10.1360
#>   -3.7759  -6.7196   0.2443 -17.3970  -9.0093   1.8225  -0.7513  -1.9545
#>   -0.8513  -8.9555  -0.9944  -0.0412  -2.0553 -15.9070  20.6484 -16.5913
#>   -0.4846  -7.7680  -5.9797 -18.5395  -7.7644  -1.0492  -3.0366  -1.5126
#>   -2.7955   1.4945  -6.2368  -5.5549  -1.0436 -22.5577 -13.3097  11.5110
#>   -1.6054  -6.9390  -2.3740  10.4213 -12.6316 -12.4618  29.1698  -7.0598
#>    9.4215   9.3437  -5.5011  11.7383   7.1842   6.2286   4.6761  21.9633
#>    7.4533   3.2306  -3.5268   0.5375   4.0495   8.6293  -8.8610 -15.2043
#>    0.8337  -4.3165   7.1755  -1.4217  -4.2569  12.2714 -11.8981   6.3360
#>    1.4984 -10.7431   3.6656  -2.7562  -0.5481  -0.4666   9.9397   4.7115
#>    1.8964  -4.1143  -4.4596   0.9131 -10.6812  16.4867  -9.9156  -2.6290
#>   -0.6854  -4.3212   2.2958   8.1793   3.2975  27.2817  -6.4204  -1.4923
#>   -2.8220   1.6065  -1.2348 -14.0992  -3.8337  -0.2564  -3.6192  -7.6158
#>    6.2940  -2.0816  11.4139  17.9484   1.0609  11.9159  10.3409  -3.1379
#>   -1.3933 -10.7616 -10.6133 -16.3385  -5.6941 -12.8734  -6.1879   1.6274
#>    0.5432  11.1469  -4.7812  -5.3665   8.7724  -2.6120   2.4939 -17.4056
#>    3.7708  -3.2995  -5.7702   8.7603   6.5936   9.7257   4.1166   9.2047
#>   -1.4285   0.2845  -3.6554 -10.6399  -9.3227  12.3171  -3.8347   5.1658
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]