Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.4685   9.1811  -6.9413  -4.6584   5.8287  -9.3106   4.1264   0.6794
#>    5.5504  -4.4867   4.9302  -0.0504   1.0971  17.4954  -3.5031  -0.6336
#>   -5.7564  -6.8771  -6.5542  -4.8754  -7.2044  -4.2765 -13.5447  -1.6377
#>   -5.6145   1.1946   9.4816 -19.9060   9.1433   3.8918  -7.1261 -12.2571
#>    6.3572   2.7419   0.2862   4.4904   2.4844   8.6092  -2.4276  10.6949
#>   12.5766  -6.3344  -0.9048  -2.3300 -10.4820   5.7454  -0.7028   2.5332
#>    1.0880  -5.1586  -1.3515   7.8267  -9.7172   6.1732  12.0964   6.7157
#>   10.0475   5.3565  -3.4849   7.2959  -1.4124   6.6867   0.4489   9.8420
#>    0.3071   4.9818   3.5782  -8.1477   1.4818  21.1880  11.7233  -0.9882
#>   -0.2324  -9.6174  16.2410   9.4230 -17.9255   2.4669  11.2504  -5.9440
#>   -0.2967   0.8283  -1.3539   0.1605   8.9076 -16.3281  -7.3451  -2.3227
#>    3.1317   1.2920  -6.1567   7.2636  -2.4825 -16.1220  -3.7493  -6.9307
#>   -1.5173   7.8981  -9.9338   5.2731   5.6987   1.9603  -0.7393   1.8262
#>   -2.9687  -6.9834   4.1400  15.3257   0.0152 -10.9046  13.7857 -11.0897
#>    2.8947  -3.2781  -4.8986  -2.7162   8.2521   0.3010   4.2349   6.6338
#>    1.1745  -0.4291   2.0160   5.2153   4.7359  10.5809  -0.1069  -6.3469
#>   -2.4331  -7.7830   2.3957  -4.7272  -7.8733  -1.7745   9.1705  -8.4665
#>   -1.4694  -3.0891  -7.1851   2.4130   6.0008 -13.3645  25.6169   6.6641
#>    2.5720  -2.2828  -3.8409   4.8690  16.6324 -10.4317  -3.4076   7.2094
#>    4.5313  -5.5293  -4.8787  13.3845 -19.0305  -6.5837   7.3865  -1.9384
#>    0.1825   2.6167  -1.2521  -0.9671 -11.6991  12.1258   8.5239  -0.4270
#>    0.7672  -2.8660   4.1785   9.4241  -7.7963  14.4731   5.5480   3.0721
#>   -7.3408   3.2802  -2.9667 -10.7473   1.2029  -7.6219   0.0825  -7.5907
#>    7.8228  -7.7580   3.1083   1.7309 -20.0681   4.6665   7.3562 -10.9278
#>    1.7902   5.5386  11.3708  -9.5959   5.1922  22.9169   6.2183  19.2347
#>    1.0059   3.7903 -14.2479  11.3911   0.0866 -14.8306   2.5477  -7.9382
#>   -5.0371  13.2370   6.5728  -8.1082  14.1268   8.2096  -8.9079   6.1505
#>   -7.5755  -6.5964   3.8890   0.8793  -8.8644  10.4535  16.9993 -22.1027
#>   -3.0355 -17.9058  -0.6813  -7.9929 -10.6610 -12.5565 -19.1345   1.8915
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]