Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  4.2338e+00  1.9037e-02  1.8047e+00  4.5128e+00 -8.5867e+00 -3.4797e+00
#>   3.8062e-01 -1.1497e+01 -1.1637e+01  5.9643e+00  1.4573e+00 -2.2887e-01
#>   2.8164e+00  4.6841e+00  1.8453e+00 -8.2536e+00  6.3480e-01  4.0345e+00
#>   3.3028e-01 -1.6769e-01  1.1061e+00  1.1459e+01 -1.1335e+01 -1.3302e+01
#>  -2.8260e+00  1.1275e+01 -1.2022e+00 -5.7197e+00 -6.3465e+00  6.6861e+00
#>  -3.0271e-01 -5.3202e-01  2.4564e+00  3.1496e-01 -8.2168e+00 -3.8754e+00
#>  -8.7447e-01 -5.1685e+00  4.0034e+00  5.8978e+00  4.0992e+00 -4.8202e+00
#>  -4.6881e+00 -4.3039e-01  1.3813e+01 -4.2605e+00  1.1237e+01  4.6943e+00
#>   3.0371e-01  5.0680e+00  2.0918e+00  7.7320e+00  1.1960e+01 -2.5651e+01
#>   3.5866e-01 -2.8851e+00 -6.6251e+00 -6.6133e+00  5.5966e+00 -3.6547e+00
#>   2.3284e+00 -4.8225e+00  7.3843e+00 -4.2651e+00  4.9250e+00 -2.1287e+01
#>  -2.3678e+00 -4.8762e+00  4.6519e+00  7.4190e+00  2.4965e-01  3.6568e+00
#>   3.0075e+00 -5.8966e+00  4.2612e+00 -1.8325e+00 -1.2830e+01  1.5268e+01
#>  -3.5632e-02 -1.0945e+00  2.7146e+00  5.6277e-01  1.5836e+00  7.5259e+00
#>   7.2388e-01  4.2605e+00  1.1958e+01  2.4617e+00 -5.3498e+00  8.2753e+00
#>   1.1326e+00 -1.1956e-01 -5.5941e+00 -1.0818e+01  1.0715e+01 -1.1137e+01
#>  -1.1029e+00  6.0325e+00  1.0477e+01  8.3775e+00  7.9551e+00  4.3034e+00
#>   4.2739e+00 -1.2414e+01 -2.9291e+00  5.1779e+00 -1.4389e+01  8.8710e+00
#>  -1.0744e+00 -1.6863e+00  2.1073e+00  9.0013e+00  6.9798e+00 -2.2103e+00
#>  -6.1973e+00  2.9860e+00  2.9549e+00 -6.1151e+00 -3.4789e+00 -9.0192e+00
#>   2.5488e+00 -3.7257e+00 -1.3156e+00  1.5459e+01 -7.2979e+00  1.8445e+00
#>   2.6741e-01 -2.4335e+00 -1.2803e+01 -1.0279e+01 -6.9152e+00 -4.2511e+00
#>   3.4768e+00 -3.1360e+00  4.2909e+00 -2.0312e+00  4.3073e+00  4.5796e+00
#>   1.7582e+00 -6.1262e+00 -1.6750e+00  7.0929e+00  7.0545e+00  9.4821e+00
#>  -5.0225e-01 -3.6602e+00 -9.5161e+00  1.1223e+01  9.7158e+00 -5.8737e+00
#>   4.1941e+00 -1.5870e+00 -5.2946e+00 -1.6462e+01  6.4701e+00 -6.0243e+00
#>   3.3760e+00 -4.2948e-01 -9.0602e+00  2.0769e+01 -1.5418e+01 -1.4481e+01
#>   1.6663e-01  6.2823e+00  1.2158e+01 -3.4432e+00  8.2631e+00  1.6999e+01
#>   7.7592e-01 -4.1670e+00 -4.2674e+00 -6.4195e+00 -1.9859e+01 -3.1348e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]