Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.0951  -4.4887   7.1706 -11.9288   2.3257   2.7172   3.0452  -3.3597
#>   -2.7929  -5.6176   3.7670 -12.7652   3.3652  -2.7576   0.4526  15.6448
#>    0.6599   7.2433  -0.3252   2.3830  12.1375   0.8236  -3.5261  -0.6931
#>    6.4398  -7.9058   3.9695  -2.3301  -3.1469   0.8938  -1.0605  -3.4158
#>    2.7213  -0.9760   2.8565  -6.9030  -3.5524   1.7784   4.2024  -3.9622
#>   -0.3493   1.6350   5.1332  -6.6857   5.0880 -10.8599   2.3775   9.2185
#>    1.4876   1.8529  13.5115  17.3941   1.3552  -0.6338   9.1118 -12.1210
#>    0.4497  -8.1962  -1.6019   6.2171 -15.4993  -0.3812  19.2781   1.8700
#>   -4.1040   9.1708  -8.9591   4.0995   9.4524   5.6654 -17.0490  -4.9553
#>   -4.3988   3.3134  17.3345   8.7962  -5.4259  -0.6027  11.6351  -6.2520
#>    4.1525   3.3320   7.6556  -3.8112  -1.1525   6.9972 -18.0031   3.9092
#>   -3.1081  -7.9507  -1.9319 -10.7696  -2.0105 -13.5473  12.6357   3.6296
#>   -1.2583  -0.2245  -2.7994  -7.6855   3.4827   7.7015   1.9749  -0.9927
#>   -1.5806   0.2801  -4.5478  12.7572  -6.0560   5.7480  10.0937  -3.4758
#>    4.4900   3.5950  -3.7303  -2.1967  12.1581  11.9122   6.0468   0.0986
#>    6.5642   0.7409  -5.2914   2.1720  -5.3922  -1.7285 -10.4751 -14.1038
#>    2.1751  -5.0633  -7.7287  17.2258  11.5906   2.6544  -0.7048   9.2895
#>   -1.3210   4.5309   3.6524  -5.4888 -14.0103 -11.4105   4.0353   6.5407
#>   -4.1806   2.2615  -4.0494   0.0864 -10.1342  -8.1789  12.2650   9.4153
#>    2.8420  -1.5807  -3.3556  14.7163  22.2583 -19.7747 -16.1991 -12.7353
#>   -2.8229   0.0763  -9.7269   0.9302  -2.6027 -24.7965   2.7590  -0.3861
#>    1.7761  -2.7763   0.2075  -6.8071  -6.1379  -5.9809   2.8131  -2.9818
#>   -1.8759  -5.5026   2.3909  -6.9407 -10.3690  10.0441   5.1503   8.0939
#>   -3.3159   6.1740  -1.2167   6.9374   0.8904   1.2617 -11.6866   0.4654
#>   -2.5251   3.1663   5.3994  -2.5233  -4.2931   7.7827  -7.2930  -0.5567
#>    2.8686   2.2017  -2.8469   8.1130   8.9044  -7.1101  -9.8535  -1.6466
#>    3.6009  -5.2135   0.9254  -3.4080   6.6523   8.6530  -6.9770  -0.4004
#>    4.0448  -3.2535  10.9099 -12.9959 -16.9925  24.0196  10.1663   0.8410
#>   -6.5253   0.0499   6.6871   6.1454   3.2574  12.9488 -21.4879  13.4215
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]