Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   9.5884   9.1867  -0.6807 -11.2495  -8.4798   9.9962 -16.1943  -7.3878
#>    4.6030   2.1766   0.0373   8.5647   5.3245  -3.8980   0.3086   0.4121
#>   -6.1647   6.3974   1.5015  -3.7941   3.9570 -13.0989   6.6442 -10.2893
#>    1.3544  -3.7661  -5.1275   3.1049  -0.8270   3.2240  -7.5382  -8.8242
#>    0.2046   2.8071   0.2395  -3.4983  -7.3622   7.9898   5.3590  10.9922
#>   -6.3524   1.3955  -0.0885  -4.2955  -4.5698 -13.8340   3.5724   1.8407
#>    0.7152   0.4493  -3.0076  16.6808   2.9304 -11.5861  10.4818   4.5856
#>   -2.5584   8.1775  -4.3524  -5.0386   6.0055 -13.1888 -14.1984   2.2009
#>    4.7704  -1.7895  -5.9199  -1.4966   1.7523  -2.3653  -7.5356   4.3336
#>   -9.7172 -10.9785  -4.8735 -12.6526   3.6146   4.0671 -11.8029   2.3457
#>    2.2558  -1.1794  -4.0693  -0.6003  -9.8439   6.0019  -0.3699   1.2649
#>   -3.5277   2.7477  18.1575  18.6541  19.0237  -5.1939  -4.1989  -0.4976
#>    1.4804  -6.0806  -9.4496  -0.2403  -6.2914   5.7766 -15.2254   0.8309
#>    0.5477  -6.1083   7.1128 -10.3014 -20.5787   2.9405  -2.6378  15.8520
#>    1.6579   4.9282  -7.8878   0.3130  -0.3566 -12.4655 -12.9049 -13.5002
#>   -0.7185  -4.9985   8.2159   6.0876  -0.0970  -5.8905   8.2793 -14.4637
#>   -4.3553   4.4211  10.5392   6.4483  -1.7880   2.7614  18.7480   6.9206
#>    7.2288   0.4167   2.2797   4.7474  -3.1379   3.6361   1.9418  15.3276
#>   -2.2795   5.5356  -1.0425 -18.1482  13.8819  16.1196   3.0586   6.4910
#>   -2.5655   2.4336  15.4982   0.7215   1.7086  10.6416   7.8823  19.0501
#>   -1.8355  -6.5471  -0.9307 -19.8658   1.0953  -2.2501  -3.3061  -7.4960
#>    3.1611   5.2546   8.2343   3.7140  -7.6445 -22.0199  -2.4940  -5.3391
#>    1.1477   2.7132  -9.8762 -15.2227   2.1685  -1.6845 -16.1989   2.2481
#>   -2.7657  -7.2634  -1.1273  -0.6223  -8.2419  -4.7483   4.5672  -5.9468
#>    1.4445  -3.8836  -3.2713   0.7211  -0.9409   0.0927  -4.5185  -6.6978
#>    6.0847   4.6487   3.6237 -10.1425  -0.0241  10.0287 -13.0792   0.6849
#>   -6.2908   4.4689  -6.2136   0.0454 -18.0304   7.1274  -8.6030 -12.6219
#>    0.7815   6.6365  -8.9672   4.0841  15.9672  11.6632   8.8963   4.2203
#>   -0.6343  -2.7324   7.3786  10.9617  -7.6065   6.5979   3.9930  -0.4208
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]