Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.9898  -7.5195   0.7766   2.2032   0.1553  -2.7225  -4.2066   2.6329
#>   -3.5387  -1.1063  -1.0588   8.9455   4.2928   2.5475  -8.5957  -6.2075
#>   -5.6727  -2.1892   2.6569  -8.5437   1.5336   2.0151  -0.3272   0.7608
#>    1.8217   5.0605  12.5835  -4.5772   1.1852   0.1242  -2.2989  -9.1448
#>    8.0308  -3.9393  -4.5214   9.0315  -5.9599 -18.5511  -9.2751   0.1649
#>    2.2616   1.0159   2.5582  -7.9131   5.2086   6.2406  -1.2246   6.5471
#>   -6.7320  -2.2519   1.6992  11.3005   5.7785  -2.9569 -12.1868  -6.2820
#>    4.4206   2.6381   1.9076   0.0742  -2.2758   4.5837  -3.9164  10.9294
#>   -1.6297  -6.4853  -2.7174  -3.6019   4.0247 -15.2723 -15.5479  -0.3213
#>   -6.3284   4.2818   1.6739  -6.7179  -0.9188  28.8853  -3.9212  -2.0959
#>   -2.1464   2.6635   4.8030  19.4628  -4.9345  -5.0830  24.4271  -3.4412
#>   -6.5569  -0.9036  -4.2747  -2.0153  -7.3221   6.9821   3.2654  -0.0204
#>   -7.0347   2.4629  10.1623   7.2106  13.7738  -7.6721   3.6484   2.9583
#>   -6.8355   4.1639  14.8304   3.3475 -18.8048  -1.3395  19.5124  -5.5520
#>   -4.9225  -5.2244  -3.4825  -0.7822  -2.2009  -5.2894  -8.3387  13.5345
#>   -3.5576   4.6229   2.8850  14.1302  -7.8738  -6.4474   7.0004  14.5603
#>   -4.4369  -8.8286  -4.4196   0.0768   1.2646   4.3639  -7.7227   1.4955
#>   -4.4141   1.2668   2.6066   5.6970   7.9733   7.8221  -4.4732  -2.0539
#>   -4.9796   2.5234   9.8420   1.6389 -23.6260  10.5197  -2.6831   8.8765
#>   -1.2655   3.0515  -1.9092  -0.5472  -6.3965  -9.5161  -0.5792 -12.6034
#>   -0.1746   2.3562   3.0861 -11.2055 -12.4839   4.7090 -12.8181   6.5696
#>    2.4227   3.8989  -7.3595  -1.1404   0.6087   7.6592 -14.1041  -5.6113
#>    0.7416   3.6094   2.4323  17.4619   0.1572  -9.5037  -0.5044  11.0575
#>    5.0645  -1.0917  -5.3143  -9.6402  -0.5751  -5.4283  -6.4432 -22.7153
#>    2.5415  -0.1628   2.5176   7.1671   9.2529  -0.2355  -1.4704   8.0743
#>   -3.9212  -9.5048  -2.9298 -10.1759  15.9667  -0.6180 -12.7656  12.6087
#>   -1.5132   7.2312 -10.8802  -7.7619  -3.8187  10.0622   8.7056 -18.6258
#>    5.9016  -0.4656  10.0182 -11.5864   4.5625  -4.1773  -8.9639  11.7874
#>    3.4899   0.3460  -4.7776 -15.4225   8.3785   2.5758   0.8950  10.0794
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]