Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.7995  -0.1451  -1.8938  -1.9651   2.4485   0.4305  -4.6356  -3.0832
#>   -3.8114  -1.9427  -2.4297  -0.8558  -2.6261  -2.4500   2.6630  16.4266
#>   -3.8863  -5.3599 -16.2798   1.0908 -10.7909  -6.2796  -0.3805   8.2256
#>    6.7751  -9.0410  -9.8385   0.1622 -12.7385  -6.1881   9.0968  -4.1427
#>    5.8584   2.6562 -11.0256  -5.1492  -5.0473  11.7467  11.7979   0.9760
#>   -0.5341   0.9938  10.1773 -12.4076   4.2016   4.0147  21.2560  16.1944
#>   -1.0517   2.6966   2.1273  -2.3470  -9.4584  -1.5511  -4.8561   0.3616
#>    4.2766  -9.9663  -0.1809   8.2637   0.0655   2.4277   3.8168  -8.0669
#>    3.5176   5.0126  -0.5013   5.0637   6.1142   3.5542  12.8272  -3.9378
#>   -1.8148   5.1640  -2.9737  -8.0751   2.1416   1.7025   5.3132  -5.4991
#>   -0.7367  -4.0171  -1.6233  13.0001   6.3936 -13.8051  -5.4729  -0.2479
#>    2.2665   3.8031  -4.2727  -0.6183   1.0269  13.5269  -1.8049   7.3663
#>   -0.1200  -1.6999   8.1082   3.9677   0.7549   8.1588   0.9257  -2.0604
#>   -0.4917   2.3832  -8.7919   3.1548  -4.5490  -9.4403   3.2412   3.8220
#>    1.3594  -6.0283   1.4802  -0.7611   1.2472   5.9034  -0.6455  -1.8716
#>   -1.0745   1.5240   8.7919   1.2158   0.9447  -5.9221  -2.8830  12.0108
#>   -7.6106 -10.1548  -8.1123   7.3686  -8.7643   9.0393   1.7774   8.4635
#>    3.1611  -0.9020  -9.2538  10.4538   8.9671   4.1494   6.3077   1.2614
#>    1.5357   2.3682   8.4742  -2.1961  -1.6700 -10.0467   4.7879  16.0799
#>    1.4845  -7.9800   7.4484   7.2253  -7.2428  10.2343  -0.8428  -0.8112
#>   -1.3112  -3.8881   8.0593  -9.8296   5.9862  -8.0760 -10.2247 -18.2372
#>    1.1315   1.8720  10.1670   7.0512  -7.5438  -5.2658  -3.8181 -11.6624
#>    0.1746  -1.2945  -9.2888   0.8866   4.9526 -28.7129 -12.4799   0.3988
#>    5.0103  -6.6799   4.8812 -15.7959   0.6490  11.8445 -14.2466   0.2270
#>   -0.2411  -6.2268 -10.0522  -4.8517  -4.2987   1.3111  -1.4363  13.7772
#>    2.7891  -0.3349  -4.5385   3.9736  -5.7397  12.7511  -4.9186  -8.6368
#>    2.4455  -4.0878 -16.3026 -17.7982 -12.0936   5.4682   3.2515  -1.6290
#>   -2.2111   2.4377  -4.1652  -9.0680  10.9755   6.4077 -11.0563   0.4510
#>    1.3298  -7.4211  -7.8205  -1.9871   4.3767   0.2401  -5.8447  11.6512
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]