Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.7189   0.4654  -0.3146   3.9441   9.1267   6.1779   4.5202   9.7629
#>   -5.9164   4.2915  -0.6774  -5.5891  14.0745   6.2017   4.7247 -11.1297
#>   -4.6001  -8.1417   6.7083   2.1118   8.7652  -3.5638   4.0507 -19.5270
#>   10.9155  -4.2621  -9.6316   3.1640  -3.5120  -3.8173  -0.3075  -2.1011
#>   -1.5232  -0.0320  -0.2707 -12.4031   1.2341   8.5350   6.6229  -0.4563
#>   -2.1924  -9.5336  -1.0534  -1.7011  10.0130  -0.7543   2.3331  -3.7918
#>   -5.7133  -3.1763  15.6879   9.6184  16.4000 -11.6381  -9.1028  -3.6933
#>   -8.3075   1.7698  -9.9447  -4.2415  -3.2254  -3.5469 -10.6890   1.0615
#>  -13.2478  -3.5342   4.0286 -10.3076   2.4403   7.1418 -15.0154  -2.9779
#>   -1.0291   2.9183   6.4252 -13.4171  12.0771 -26.6983  17.0866  -0.2578
#>    9.3938   1.9358   0.7177  15.4144  -2.0149  -7.0842   8.4709  -7.1891
#>    1.2034   3.8274  -6.5285  -7.4459   3.0881  18.7283 -18.6641  -3.7251
#>    3.0244   7.7607  -1.3843 -25.6028   5.6696  29.5993   5.4250 -12.2513
#>   11.6302  10.0917  -0.8680  -6.6427   3.7809   1.7397   4.7254  12.3179
#>   -4.5016  -8.8641   4.1427   4.1620   7.8718  -8.1157  -8.3436  -7.6961
#>    1.2771   5.8871   0.1172   4.3974  -6.4844   3.8681   4.2868  -0.6329
#>   -3.2202  -7.8925  13.4067   5.4161  10.0722   5.0142 -11.0770 -15.6162
#>    2.4781  12.3500  -9.0077   6.1409  -8.3381   2.7441  -8.1381   6.8781
#>   -2.5296  -6.3919  -1.9955 -13.7438  -2.6014  -6.1622  13.1916  -1.5824
#>   -2.3142   1.3868  -2.6371  11.9843 -20.4733  -0.6383  -0.3163   2.3097
#>   -3.5691  -0.1326  -4.3173  -5.2492 -12.2591   9.5318   4.0171  10.1427
#>    1.6368  -3.0935  -3.7063   4.0719  -6.5891  -6.1968   3.6046  -6.1577
#>    4.2860  -0.2571  -1.4456  17.2456   2.7306  -1.3091  -3.1286  -4.5182
#>   -0.3486   1.4387  -7.2499  10.6834  -4.5659   2.8581  -3.5703  -7.4222
#>    1.6656   6.5691   8.2412   4.6968   5.5554  -6.5881  -7.8120 -16.0592
#>    3.6165   6.7479  -7.1273   2.3360 -23.5106 -11.4830 -10.3809  23.6060
#>    1.8733   3.1452  -9.1464   0.9714 -12.5836  -2.9398  -5.1949   1.4322
#>    5.0954  -4.0262  -3.2574  13.5642  -5.7637  -5.8741   1.0629   7.9760
#>    1.9679  -3.2147  -0.4391  -3.6081  10.4532  -6.5897 -11.8092   9.7918
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]