Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.2874   2.1602   1.2526  -3.3547   6.3589  13.4857 -12.4803   8.3072
#>    5.9402  -0.8409  11.2891  -6.2280   2.0193   9.6301   5.6913  -8.6700
#>   -4.7374  -2.3570   3.1580   7.7398 -23.7808 -19.2438  -2.1502  14.3772
#>    0.6523  -0.4368 -13.2043   7.1822  -4.8584 -10.0034  -6.1820   9.1084
#>    1.2427  -1.8380  -7.2037  -4.1628   0.4724   8.7805 -11.5545  -2.1589
#>   -1.9563   7.5760   9.1303  12.1277   7.2203   9.8649  -2.6101  26.0164
#>    0.5206  -0.3507   3.7937  -7.1877   4.4094   0.3254  11.1104  28.2837
#>    0.2210  -1.1287 -15.1017   0.1891  -3.3935   1.0815  17.6563  10.1911
#>   -0.5243  -7.6844   7.7109   6.9407  -3.4129 -12.0396   1.6615  -2.0911
#>    0.5141  12.6622  -4.1989   8.7871  -4.0435  20.2766  18.4881   5.1692
#>    0.1230  -1.2851  -4.2556  -3.5830  -5.3946  -2.3487  25.9456   5.0268
#>    3.0123  -8.6552   4.9295  -0.3869   0.1757   5.0003  -3.5342  -6.6768
#>   -2.0143   2.5505  -1.2122   2.7036   3.5080   0.1396  -6.3571  -4.9591
#>   -0.6604   6.3349   4.8487  10.5876  16.1622 -14.6632  -4.4367   6.3069
#>   -1.6987   0.3463   4.8421  -0.2171   4.0974   4.9722  -7.8555   7.2371
#>   -2.3939  -4.1757  -3.3780  -7.0899  -6.8553   4.3105  -2.7587  -8.4139
#>   -0.1458  -5.2668  -5.5106  -3.7528   9.8936   9.7397  -9.7447  -0.3929
#>    0.1574  -6.4607   8.3319  -0.2524  15.9423  -5.1316 -10.9248  -8.6726
#>    4.9185  -9.7123  -2.3832   0.4603   4.2285 -13.1005   2.8222  -3.7271
#>    4.9020  -5.1043  -4.0577  -2.4182   6.4507  26.3785   7.5186 -24.7669
#>    0.7573   0.9935   6.9451  -4.4153   7.5132  -6.6906 -12.9394   4.4447
#>    3.4864  10.7350   6.0699  10.0592  -7.8695   0.2528  15.7802  -4.9364
#>   -0.4667   5.9711   3.1923 -12.4438  -9.2217 -24.3899   1.9654  22.5933
#>    1.0739   5.1420  13.6287   0.0556   1.0393   3.7035  -5.4045  13.7047
#>    3.7065  -0.9727   2.0113   2.2002  -3.3668  15.1920  -2.3484   7.7476
#>    0.5932   7.1558   6.1206  -6.9478   8.8130  -5.8559  -3.9549   3.0421
#>   -2.7964  -0.3328  12.9517  -0.7100  18.1273  -1.2782  -0.1395   4.7146
#>   -1.1306  13.4673  -3.2738   1.3099   9.6073  21.3842   4.9132   0.7214
#>    1.6620   0.3419   0.4320  -2.7167  26.5992  10.2905   1.2853 -14.1951
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]