Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.9485  -7.5550   9.0444  -4.0951  -4.1640  -8.8162  12.0670   1.4320
#>   -3.0729   1.8121   0.0707   2.4596   1.4147 -12.5996   7.7892  -5.7023
#>   -2.5299  -0.6959  -2.7286 -21.7467 -10.0325  15.1177 -13.3289  -5.6568
#>   -3.5884  -2.1980   2.9200   9.7079  -9.2078  -4.1193  20.5217 -14.2474
#>   -4.8212   6.0793   1.0434   6.2387  -3.0315  -1.2627  -6.6053  -9.9331
#>   -1.1358   0.5861 -13.4190   6.3464  14.1987 -10.8189 -11.0309  -2.8002
#>   -0.8164  -0.3891  -1.1462  -8.8261  -6.2091 -15.7911 -11.5470   1.7427
#>    3.1875 -10.2459  -9.0321  15.3996  -1.2204  -6.0075  -2.5732 -10.1810
#>    3.2754  -1.1082   3.5179  18.5654   8.0126  -7.1747  -6.0657   4.8946
#>   -6.7334  -6.8606  -2.9777  -8.3373   5.0099  -2.9725  -3.2855   8.2471
#>    2.9396 -12.5519   1.6655  -8.1706  -2.2176   5.0109  -3.6025  20.4398
#>   -3.2390  13.3909  -2.3776  -3.9026  -9.2668   5.4534   4.5531  -1.5744
#>   -2.1506   5.0748   3.0234  16.4141   4.3317 -13.4767   5.7470  11.8165
#>   -1.5072   1.8530  -5.7444   8.4292  -0.8130   3.6022   9.8223  13.1478
#>   -3.7704   0.1722  -5.6879   8.9908   4.4606  -1.5726  -8.7933   8.8696
#>   -0.5834  -2.9866   4.2465 -16.1334  -1.2829  -0.5182   1.5126 -13.9695
#>    2.1182 -14.3786  -1.3419   3.6074   9.8502   1.6422   6.1153  10.5370
#>    3.3439   3.5070 -11.9443   4.8237  -7.0056  -9.5203   3.9175   1.3672
#>    1.1024   5.8756  -4.5663   7.0922  15.7058 -10.8411  -5.7436   4.2023
#>    4.3646  -5.3318  -1.1575   3.4429  -8.7055  14.7078 -10.0577  -0.2754
#>   -0.6841   9.0061   0.5627   4.9533   4.4213   0.9984  -7.5703  -7.2945
#>   -2.4085  -2.9900  -2.7380  -5.6239  -1.1478  -5.2871   0.3699  -4.5467
#>   -1.1103  -6.3872  14.5306   2.2437 -13.2758  12.4493  -8.1787 -10.0135
#>    3.2884   0.5758   3.3570 -14.5938  -9.6044  -7.8061  -5.6873 -11.7722
#>    0.7775  -5.2431   5.3999  -0.4828  -8.9769  -3.0376   8.4991  -7.7100
#>   -3.0194  -2.6538  -3.6716  -2.3849 -14.1505  10.5691  -7.8289  -7.2762
#>   -6.0732   1.8255   5.4464   1.6714   3.3393  -3.3272  -0.9127   9.1583
#>    1.6341  -6.1641  -2.8389  -4.9926  12.9236 -15.4878  -0.4643  16.9894
#>    0.7260   3.3509   3.2419  -4.1139  -2.8246 -10.1746  20.2263  10.9091
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]