Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  7.5953e-01  7.0703e-01  8.4139e-01 -1.6894e+00  4.0227e+00 -2.6119e+00
#>   2.6907e+00 -6.2970e+00 -6.7349e+00  2.8829e+00 -9.0819e+00 -1.1938e-01
#>  -1.2385e+00 -3.9971e+00 -6.5574e+00 -1.1957e+00 -1.1293e+00 -6.8133e+00
#>  -3.7120e-01 -3.8545e+00  2.1281e+00 -6.8125e+00  1.8109e+00 -1.0898e+00
#>   5.7846e+00 -3.7728e+00 -8.0982e-01  2.2299e+00 -8.5520e+00  1.2505e+01
#>  -2.3103e+00  9.1139e-01 -9.2152e+00  1.8412e+00 -8.2403e+00 -7.9738e+00
#>  -7.0113e-01 -6.8272e+00 -2.0267e-01  9.0337e+00  1.1774e+01  6.1642e+00
#>  -8.5514e-01 -8.0758e-01  4.6574e+00  3.6173e+00 -1.4823e+01  3.4994e+00
#>  -6.0144e+00 -1.0595e+00  7.8407e+00 -9.7874e-01  7.9307e+00 -5.0101e+00
#>  -6.3578e+00  7.1297e+00  7.6885e+00 -7.8381e+00 -5.0227e+00 -8.2524e+00
#>   2.9732e+00 -7.8615e+00  6.4104e+00  2.1664e+00 -1.5703e+01  1.0529e+01
#>   3.2328e+00 -5.6867e+00  2.9833e+00 -4.5000e+00 -8.9110e+00 -3.9557e+00
#>   4.9048e+00  2.3922e+00  5.2938e+00 -4.7995e+00 -2.3921e+00 -7.2680e+00
#>  -4.5844e-01  3.0707e+00  5.0348e-01  4.7116e+00  4.8040e+00 -1.8241e+01
#>  -1.1477e+00 -2.0127e+00 -1.0350e+01  1.3217e-01 -8.7087e+00  5.8197e+00
#>  -1.9813e+00 -1.1315e+00 -1.0570e+01  2.3607e+00  1.7524e+00  4.0657e+00
#>   2.0564e+00  7.6804e+00  7.1764e+00 -7.4391e+00 -3.0052e+00 -1.1390e+01
#>  -3.1207e+00 -2.9382e+00  3.9696e+00 -7.5983e-01 -5.2182e+00  5.6373e-01
#>   8.8749e-01 -7.6767e+00  1.6628e+00  1.9530e+01  8.7764e+00  1.2558e+00
#>  -4.9246e-01  3.7833e+00 -2.2112e+01 -1.1902e+00  1.8893e+01  4.6643e+00
#>   4.2482e+00  2.7933e+00 -6.1990e+00  2.2185e+00  7.2540e+00  6.1754e+00
#>   9.1364e-01  6.8322e-01 -1.0584e+01  2.3514e+00  5.2057e+00  8.7360e+00
#>  -1.6657e+00  1.3665e+00  3.8860e+00  1.9474e+00  1.7095e+01 -1.0886e+01
#>  -1.4666e+00  2.2988e+00 -5.8078e+00  6.5675e+00  9.0495e-01 -8.1416e+00
#>   1.5038e+00 -2.7258e+00 -5.6332e+00  5.7973e+00  1.0291e+01  8.4069e+00
#>  -4.5248e+00  3.4989e+00 -7.0232e+00 -1.2628e+00  1.1366e+01  5.3985e+00
#>   1.6114e+00  4.5156e+00  1.3252e+00 -4.3028e+00 -2.7342e+00 -9.3208e+00
#>   1.5332e+00  1.8359e+00 -4.1384e+00 -1.9701e+00  1.0329e+00  1.1907e+00
#>   1.9063e+00  1.9982e+00 -9.8923e+00  1.6339e+00 -1.0022e+01 -6.8436e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]