Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.9159  -5.1540  -8.6407   1.4169   7.7723  -0.9074  -1.2768  -2.6646
#>    2.7850  -0.7556   7.3809  -8.6329  -2.4106   6.4078   5.8325 -14.4589
#>    8.5966  -4.6427  15.4770  14.1256  14.0556  10.6746   7.2098  -8.5990
#>   -1.7034  -3.1659   2.7919  14.2010   9.2035   2.9884   3.9466  11.6345
#>   -7.0826  -4.4565   1.7484   9.9009  -0.2955  11.1833  -5.7876 -15.5914
#>    4.4867   0.3309   8.6406   7.9670   5.0275 -12.4850   3.2812  -2.8273
#>   -2.5427   4.2344   5.8830  -9.1184 -10.0951 -20.8630   0.6038   0.9873
#>   -0.2993   3.1755   5.3570   3.7555   2.8905   8.3428  -5.7991   5.6459
#>   -4.6348  -0.5076  -1.7182  14.8867  -4.7187  -8.7491   3.4125  -7.5658
#>    0.1968   2.2815  -2.5736   0.5207  -5.4746   5.0087  -1.5563  11.0938
#>   -1.9960   2.4597   2.3965   5.3210  -0.3385   3.9406  -6.3698  15.7144
#>   -0.9094   0.0837   0.3484  -4.8033  -7.6743   7.0276  -0.9550  -0.1992
#>    6.9913   4.9084   1.8998   5.5591   7.4370   5.4938  -3.3164  -3.9833
#>   -5.8261   4.2586  -8.8793  -6.0372  -3.9463  12.5458   3.2819  -1.8850
#>    0.2370  -1.6074  -8.4985 -11.7366  10.7784   4.2156 -14.4694   0.5764
#>   -1.0972  -1.4951   1.9186   1.4770   6.5715 -10.7129   5.5953 -21.3106
#>   -3.7381   6.1374   8.2005  -3.5257  -1.4487   0.2646  -0.7516   9.8883
#>    1.5068  -2.0193   4.2033   4.1193  13.8637  10.4383  12.5359  -0.9208
#>   -0.1664   6.5587   0.5481   2.8541   8.5529  16.2174  18.5806   8.5356
#>   -7.0650   2.8919  -9.1629   1.5382   8.5688  11.2306   2.6448  -8.6566
#>   -2.7598  -3.5010 -10.7745   0.3279  -7.9446  -4.2581 -10.1226   3.9758
#>    3.7423   5.4416   0.4367   7.0222  -7.2946   0.4530  -1.3182  -0.0099
#>    0.7404   1.5203   2.9632   3.0783   7.4996   5.6840   5.8306  -9.2904
#>   -2.7931   4.2825  -5.6544   4.2367  -2.1556   9.5286  -5.3502   2.2850
#>    3.4336   3.4484  -5.9310  -1.6021   3.8960   2.3783  -3.9539   0.4996
#>    2.1761  -3.0516  -1.4590 -11.8699  10.6404  -9.4690   7.4111   0.5877
#>   -3.9002 -12.2363  -6.4090   0.3509  -0.7467   6.9245   6.9181  -3.8502
#>   -1.7052 -13.2799   1.3924   3.3538   5.1041   6.0895  14.7621  -0.9162
#>    6.5656  -8.8159 -12.9885   6.5001  -0.0142  -0.6982  27.4004   7.7760
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]