Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.3112  -2.1196   7.5942  -2.7915  11.8265  -8.3590  12.4819  -0.6369
#>    0.0802   1.3795   2.4840  12.5892  -0.7745   5.3532  11.4117  -1.3862
#>    3.8194   9.4039   8.9577  -1.8941  -1.7651   1.6859 -23.0677   8.9273
#>   -0.3154  -7.4152  -2.3606   3.7388  -3.9438   0.9044   0.0083   2.8449
#>   -4.5367  -5.9473  -0.8035   6.5663 -12.7920  -8.9248  -6.2699  -6.1614
#>    0.0923  -7.8105  -8.1106   9.4216  13.3656  -5.2258  10.0933   5.1080
#>    3.7603  -4.1590  -4.0168   6.3243   4.4002  18.6522   3.0183 -13.7125
#>    0.8682   1.2732  10.7164  -4.7349  -1.1748   1.3088   0.3591   1.6680
#>   -7.4220  -6.1132   3.8336  -0.8787   6.4239 -12.0123  -2.7239   7.6386
#>   -5.7212  -7.7390  -8.1579  16.4139   4.3584 -11.4493  18.2354  -9.5933
#>   -1.4068  -1.7251  -3.3839  -7.1890   5.6913   8.6660   1.8879   7.9529
#>   -8.6306  -2.5689   9.4840  -1.5062   4.1779   9.2571  -6.0533   0.3875
#>    2.3298  -4.9902  -3.4189   6.3549  14.2190   4.6703 -13.2181   8.1812
#>   -1.1689   9.6523   3.8674   8.3082  -2.6397   6.2130  -3.2542  -2.7572
#>    1.8344   5.1045   4.5737  -0.5591   7.4634  -5.1236 -11.1039   2.7034
#>    0.5110   3.9990   7.1623   4.6484   3.6263   9.2962  -0.5751  -4.8554
#>   -3.8653  -7.4075  -0.4823  -4.6069 -10.5740   8.3344  -3.5220   2.0902
#>    5.7178   0.5638  -2.1280 -16.6050  -4.8491  -3.8124  -1.1205 -10.8786
#>   -5.3356   6.3846   1.3749 -14.2427  -6.4284  -6.4704   0.7593  -2.0552
#>    0.4396  -1.6952   0.2797 -15.0990  11.9879  -2.2374  -2.8209   2.4761
#>   -3.0175  -9.6659  -3.7634  -8.5892   8.5124  -8.9253  10.3725   7.6230
#>    3.5931   8.9307   0.0323   6.2264 -12.4956 -12.2949 -10.4104  -6.1367
#>    0.1817  -2.3011   1.0604  -8.2184 -11.4619  12.3151  -2.2631   1.0789
#>    0.1980   0.3979  -6.8784  -7.0583  -7.5297  -8.3582  -5.1671  -8.7593
#>   -2.9895   2.3803  -1.1402   4.1413  -5.6800   6.5754   0.4876   7.2792
#>   -7.4429  -3.8539  -6.6515  13.5213  10.8448 -14.7325  -1.9881  -6.1415
#>   10.0384  -5.3269   5.5979 -22.2332  -7.9198 -19.1819  -5.0183 -10.6949
#>    3.6605   3.9902  -3.6392  -2.3762   1.7894   0.3500  10.3279  -7.0732
#>    1.3340   6.3783  -5.8503  14.3427  -5.1110 -23.8275   1.3495  -2.1176
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]