Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  2.4365e-01 -2.9721e+00  6.8480e+00 -1.6373e+00  2.3769e+00  5.6635e+00
#>  -4.7979e-01 -6.3245e+00  2.3380e+00  2.4531e+00  2.1324e+00 -3.7405e+00
#>   1.1657e+00  8.2607e-01 -1.0680e-01  7.8207e-01  9.7037e+00  1.0192e+00
#>  -3.0710e+00  1.1264e+00 -1.5942e+00  6.2803e+00  6.1353e+00  4.5738e+00
#>  -6.2006e+00 -4.3330e+00  1.0519e+01  1.4317e+01 -2.4329e+00  6.9640e+00
#>   6.9286e+00  4.2712e+00  6.2659e-01  7.9694e+00  4.4722e+00  1.1182e+01
#>   3.4881e+00  5.1517e+00  6.6096e+00 -1.4186e+01  1.0529e+01 -4.8561e+00
#>   6.3199e+00  2.4654e+00 -3.8422e+00 -5.6996e+00  4.9116e+00 -1.6662e+01
#>   3.7122e-01  4.3614e+00  1.0689e+01  3.4911e+00  9.1354e-01 -8.0560e+00
#>  -3.7942e+00 -2.6215e+00 -5.3738e+00  2.2192e+00  1.0772e-01  8.9091e-01
#>   8.1529e+00  2.9559e+00  1.0865e+01 -1.0410e+01  5.9542e+00  1.1013e+01
#>   2.6979e+00  1.3461e-03 -3.3795e+00 -5.8293e+00  2.8016e+00 -7.3702e+00
#>   9.1957e-02 -5.2603e+00  4.5235e+00 -5.0339e+00 -1.6594e+00 -9.9995e+00
#>  -8.5298e+00  1.6388e+00  6.9675e-01  4.3034e+00 -7.7983e+00  2.6215e+00
#>   6.7891e+00 -1.4353e+00  3.6641e+00 -1.4874e+00  1.1525e+01  7.5028e+00
#>   2.9892e+00  1.6542e+00 -6.2574e+00 -4.5632e+00 -2.5922e-01 -3.2924e+00
#>  -5.5989e-01  9.8481e-01 -8.4391e+00 -5.4168e+00  1.0593e-01 -9.2548e+00
#>   8.1973e-01  4.3895e+00  7.4694e+00  1.0044e+01 -1.1875e+01 -1.5783e+01
#>  -6.3888e+00 -9.9130e+00 -6.1612e+00 -7.6215e+00 -2.8224e+00 -5.7185e+00
#>  -1.7693e+00  3.0617e+00  1.2734e+00  1.5337e+01  5.8143e+00  1.6637e+01
#>  -3.3520e-03  4.0348e+00  1.4197e+00 -8.3225e+00  1.7483e+01  6.9660e+00
#>  -1.0846e+00  1.3121e+00  4.0062e+00  3.3753e+00  2.6189e+00 -2.0771e+00
#>  -3.2307e+00 -3.1215e-01  1.6519e+01 -3.6729e+00  4.3811e+00 -1.1469e+01
#>  -5.8474e-01  1.1425e-01 -1.4182e+00 -1.4947e+01  5.4099e-01 -1.0068e+01
#>  -6.0056e+00  1.4064e-01 -6.4092e-01  4.9008e+00 -3.7999e+00 -1.4268e+01
#>  -2.2661e+00  4.5907e+00 -1.2622e+00  7.6245e+00  1.3677e+01 -1.2660e+01
#>   9.3026e-01 -6.2409e+00  1.8616e+00 -2.4255e+00 -3.5482e+00  1.1179e+01
#>   8.8858e-01  1.8426e-01  1.4079e+01  8.1587e+00  5.7919e+00 -1.0599e+01
#>   6.1588e+00  5.7552e+00 -3.0679e-01  6.8251e+00  2.0030e+00 -3.7346e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]