Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.9557  -6.2392  -6.9974  -6.9928 -12.5931  -5.8640 -12.4860   8.1423
#>    3.9160   0.8717  -6.8649  -3.2214   4.0335  -3.9126   7.3392  -9.7711
#>   -2.0386  -1.5058   2.5549   7.7955   1.8172  -6.7863  -1.0154 -15.5592
#>    0.2799   4.0606   1.2039   3.4876  16.2842 -14.0402  -2.4651   5.2014
#>   -2.1798   2.6668  -9.6360  -0.6237  -8.2889  -5.2351   2.9309 -26.2994
#>    0.7757  -2.8128   5.4106   1.0860  -6.7512   3.1330  17.0394  -5.9895
#>    4.7390  -1.5558  -1.8776  -1.2904   1.5104   4.4434  -6.2199   4.6927
#>   -2.2311  -7.0393   7.7515 -13.1909  -0.7639   1.0983   9.4670   9.1284
#>   -1.1036  -6.5822   4.1991   6.4045   0.3243  -3.8063  16.3235  -9.8845
#>    1.1542   3.1456  -0.5486   0.9143 -10.0254  -7.3819  -8.9610  -8.7329
#>    6.6034   3.5090   2.7321   0.1073  -7.4768  -0.5121 -10.6512  14.4129
#>   -3.6725   5.0238   2.8656  -1.8660   3.2309  -3.3422   0.2021 -10.7565
#>   -1.0511  -8.2084  -0.7439  -9.1868  -2.5816   5.9544   3.1274 -10.1550
#>    5.2564  -6.6627   0.2840  -8.2406   1.2869   6.6108  -5.5033   9.3111
#>    5.6738  -3.2030   9.4850  -5.2457   2.1743   0.7151  13.6999   0.0026
#>   -1.3491   1.3042  -1.0155   5.8480   6.4847   6.2559   6.8746  -2.4479
#>    3.3861   5.7923   9.0212  20.6153  -1.5906  -6.4923   9.4587  -8.0838
#>    7.1973   2.0348   0.8999  15.0214   9.1641   7.5499   6.3349   1.0657
#>  -13.2429  10.5906  -3.9765   0.4404   7.4360  -5.7972  14.5966  -7.8735
#>   -2.8900   5.2601   6.6940   9.7277   3.9613   0.8896   4.3650 -19.8172
#>    0.9261   1.9587  -2.9414   3.5663   1.2206  -0.2715  24.7229  -3.8031
#>   -2.1170   0.4828   7.0732   4.5047  13.3673  -0.7821  15.5842 -18.1054
#>    3.5200  10.9799  -4.5269   6.0489   2.6578  -8.0035  16.4604 -10.4719
#>   -5.7077  -2.8587   1.1433  11.1610   2.2595  -0.9442  -2.3371  -0.1792
#>   -1.8444  -2.8167   1.2754   3.9691  -7.6685 -19.9907  16.2202  -0.3755
#>    2.5707  -3.9558 -11.5116  10.4353  -8.7170  -8.2050  -9.6043  -0.6599
#>   -6.5755   9.3077   0.9388   2.9876   5.6337  -3.1834 -24.4450  -1.3222
#>   -4.6272  -5.0934  11.6718  -7.4212   9.0785   8.5756  -2.9621   0.3496
#>   -1.0230 -19.8325   3.7951 -13.1347  10.7991  -9.7941  18.9331  -7.3413
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]