Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.3035   0.4481  -5.0596 -11.8292  10.8426  -9.4295   7.0128   3.7800
#>    0.4403  -2.0022   6.4000  -6.7104  12.8239  -0.4372  -5.9707   2.1858
#>   -1.3720   3.4794 -13.5963  11.9814   5.3001   4.1081  -3.3605  -8.7934
#>   -4.7407  -8.8377  12.0792  -4.0575  -6.8743  13.4062   4.8329  11.0056
#>    5.0320   0.3792  -3.5476  -2.8391  -3.1429  -7.6752   0.6302   5.1932
#>   -4.7169   3.4536  -7.2866  18.2834  -6.0217  -5.8036 -12.8084   1.0172
#>    4.5623   3.7803  -1.4346  -2.4375  -4.3329  11.9505  -8.6601  -5.6470
#>   -5.6875   1.8887  -0.2950  -9.8946  13.5670  -4.6727   8.8606  -0.7320
#>    1.1476  -0.6409   7.4686  -4.8978   4.9976   2.2008   3.5378   4.4474
#>   -4.3790   6.4687   1.2181  13.8896  -3.3211  -1.0496  -3.7257  -1.4200
#>    2.2310  12.5511  -9.9625   1.7718   4.7927   4.8002 -10.3118  -1.7304
#>    2.5731  -8.6846   5.0465  -9.4715  12.1452  -3.2591  -2.8932 -15.3467
#>   -3.7502  -0.6489  -4.4274   6.3947  -2.8830  12.3230 -15.3760   6.4014
#>    0.6963  -9.7238  -0.4151  -6.6399  11.9392  -4.1924  -4.4927   1.7733
#>    3.9864  -4.2329   7.7446  -7.9324   0.7641   6.4140  -4.0364   5.5168
#>   -4.1890  -2.7953  -1.8892  18.0011   5.4890   6.4764  -6.0323 -11.9552
#>   -3.3385  -6.9691   3.9089 -10.8768   1.2664  -3.7267   5.9616   8.5508
#>    2.9317  -1.3152 -10.6889   3.1548   0.0569   3.0823  -6.6851  -1.4608
#>    2.1433  -1.4668  -2.0803  -8.5710   8.5239  -1.8897  -3.6180  -5.5814
#>    1.6904  -5.1741   7.1962  -2.4688 -20.3685 -12.9950  -0.2711  19.2038
#>   -6.1345  -1.2578  -4.6173  10.5142   5.2856   2.9202  -9.4572  -0.2902
#>    8.2417  -0.1307  -4.4208 -19.6616  16.9986  -7.8913  -2.3882  -4.6841
#>    3.4336   4.2103  -9.8411  13.4427  -1.5484  -4.7570   2.1481  -3.2764
#>   -1.5849   2.1612   7.7834   5.2074 -10.3073   1.8559   3.5157   5.3976
#>    7.2546   2.0638  -7.1938  -3.0591 -15.3192   7.7636  -1.3501  -4.6437
#>   -1.5400  -3.8867  -4.9012  14.6327  13.1751   5.5784 -13.3179   5.5128
#>   -2.5933   7.8274   8.3083  10.0600   4.9313  -7.2610  -6.5542  10.5603
#>    0.0827   4.2823   5.0638   0.1798  -0.5720   2.3141  19.4882  -3.2472
#>    2.0680  -2.5997   7.1024  12.3649   3.9101  -9.5615  -3.6811  -7.7470
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]