Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -4.4843e+00  2.3130e+00  3.4133e-01 -3.5520e+00 -4.4295e+00  5.5240e+00
#>   3.1959e+00 -8.1423e+00  3.0148e+00  1.1008e+01  1.4812e+01 -1.8168e+00
#>  -8.9596e-01  8.0789e+00  1.7795e+00 -2.8513e+00  6.6279e+00  7.7707e+00
#>   9.7653e-01 -6.5055e+00  5.4519e+00 -1.3046e+01 -9.3874e+00 -7.9701e+00
#>   2.9504e+00  1.8147e+00  6.2318e-01 -1.5712e+01  1.6675e+01 -6.1861e+00
#>   3.7141e+00 -2.0715e+00  2.2523e+00 -8.2909e+00 -7.3146e+00  2.5318e+00
#>  -2.5745e+00 -3.3051e+00 -3.4884e+00  1.1561e+00 -5.8482e+00  9.3024e+00
#>  -8.2847e-01 -1.2266e+00 -7.9974e+00  5.1846e+00  9.9515e+00  8.1160e+00
#>   8.8668e+00 -7.3124e+00  9.2506e+00  9.7575e+00  1.4347e+01 -3.0325e+00
#>  -4.8020e+00  2.0573e+00 -4.6270e+00 -7.4152e+00  1.6612e+01 -3.2806e+00
#>  -7.7793e-01  7.5017e+00  7.7050e+00  3.5474e-01  6.4844e+00  3.3894e+00
#>   4.4851e-01 -7.7239e+00  6.2528e+00 -7.4626e+00  1.9190e+00 -1.2318e+01
#>  -4.6392e+00 -1.0852e+00  6.5091e-01 -4.1212e+00  5.5677e+00  1.0243e+01
#>  -9.7471e-01 -6.1318e-01 -2.5796e+00  1.9290e+00 -1.1198e+01 -4.0838e-01
#>   3.7467e+00  2.3092e+00 -7.7212e+00 -8.0049e-01  1.1201e+01  1.3715e+01
#>  -2.6419e+00  4.1969e+00 -8.3271e+00 -3.3454e+00  1.8238e+01  1.1890e+00
#>  -6.4129e+00  1.9140e+01 -9.6464e+00  8.8715e+00  5.4205e+00  1.3130e+00
#>   2.3983e+00  1.3030e+00  5.7785e+00 -8.4749e+00  1.5447e+01 -4.2593e+00
#>  -3.6665e+00  1.2102e-01  1.0999e+01  7.1253e+00 -4.6136e+00 -1.4220e+01
#>   1.3796e+00 -4.2887e+00  7.4378e+00 -1.0243e+00 -4.5362e+00 -1.2945e+01
#>  -4.2184e-01  4.3376e+00  9.7991e+00 -7.8944e+00 -8.2710e+00  7.3169e+00
#>  -5.0907e+00  1.1347e+00 -2.0688e-01 -2.0048e-01  2.3178e+00 -1.1658e+01
#>  -3.0281e+00  2.6979e-01 -1.5355e+00  1.7114e+00  3.6714e+00  1.3378e+00
#>   7.9670e-01  3.1156e+00 -5.5000e+00  1.3588e+00  2.6656e+00  1.2629e+01
#>   3.3852e+00 -8.5875e+00 -9.8434e-01  4.1453e+00  2.0130e-01  1.0675e+01
#>  -7.8654e+00 -3.1136e+00 -2.9503e+00 -7.0298e+00  1.0532e+01 -3.1287e+00
#>  -1.4350e+00  2.7620e+00  2.5055e+00 -2.6649e+00  7.9525e+00  7.9308e+00
#>   6.4318e+00 -5.1500e+00  1.0942e+01  7.8325e-01  5.7141e+00 -9.0563e+00
#>  -1.0834e-01 -7.4162e+00  1.9310e+00 -5.6165e+00  5.9194e+00 -1.9168e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]