Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.0506   7.4688  -0.0877   0.5742   8.8656  -1.3745  -8.1519  -6.9962
#>   -1.9839   0.5441   0.3912  -7.4492   2.5189 -11.0793  13.6957 -11.4519
#>   -2.1856   1.1128  -6.7692  -3.0691  13.1768  -1.5946  18.3523 -18.6071
#>    1.8359   2.9372  -2.2482  -1.6515  15.4322 -13.5834   6.4129  -4.8568
#>    1.5974   0.0680   0.8345   5.4423  -4.4447  -0.7605   2.0426  -9.7022
#>    3.4704   6.9456  -4.1573  -1.2230 -22.4217  11.0181  -0.0671   5.3506
#>    5.7123  -1.9331 -11.4807  -8.3300  -3.1865  13.2485  -2.4772 -10.3330
#>   -1.7020  -1.4650 -17.0865   3.8190  -3.9717  -4.0679   1.1000  -5.2525
#>    1.9103  -6.5222  12.4804  -8.1322  13.8380  -3.8399   4.2926  -2.5629
#>   -0.7634   4.0369  13.0089  -0.2926 -14.6063   8.6548   5.3262   1.2563
#>   -1.4625  -0.3283 -13.4046  -3.1667   2.5641   5.8352  -1.3870  10.3311
#>    0.6642  -6.7903  -4.1160  11.3402   1.4077  -5.1968   3.9631   0.7013
#>    0.3741 -11.8573  -2.1518  -1.7755   1.4357   0.7814   9.9179   3.8877
#>   -3.1310  -4.5805   0.3219  -1.7130  10.6109 -10.5513   7.2368  -7.5752
#>    0.4829  -0.7932  -1.0479  -7.6880  -5.9639  -7.7365  -3.6186  -3.7121
#>   -3.8586  -4.1306   4.5743  -7.5741   9.2638   0.3432   5.6352  -7.4261
#>   -1.3536   0.6279   4.5215   1.4695   1.5553   2.8744   0.7270  -8.2396
#>   -2.4308   5.1688  13.4067  14.9458  -1.9480  -8.3040 -29.8605   2.5275
#>   -1.8526  -1.2218  -0.7697   3.5035  10.1194 -10.5041  -0.8653   5.5003
#>   -1.0064  -5.2628   0.2168   4.7964  -4.4932   5.1786   8.3515  -8.5059
#>   -1.8420  -3.2435   1.5433   1.6402   4.4952   3.4270   2.0898  -0.3791
#>    1.5133   0.4380  -3.5040  -1.4340   0.2612 -16.0531   7.5022   9.6037
#>    5.8235   0.7934  -2.9553  -6.4450  10.3128  -1.0250 -12.7137   6.8607
#>    4.1342  -1.1640  -4.4914  13.1564 -13.5098   5.9334   5.4582   6.9176
#>    2.5396  11.4854   2.1042  21.8469  -3.8805  18.0645 -12.7992  -0.7253
#>   -3.3337  -0.0561 -13.8665   5.6583  10.1765   8.5753   2.7939  -3.7829
#>    3.9784  -0.8639   7.4228  -0.7537  12.8171   4.5340   6.9405   0.9622
#>    0.0817  -2.3672  -8.4093  -2.0992   2.6703   5.9387   5.8021  -4.0096
#>   -2.8070  -5.3752   1.2818   5.9470   3.5821   4.2418  -3.0354 -12.0863
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]