Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -6.8145   3.9718  -4.1707   6.8746  -3.7859   9.0100  -0.8963  10.1136
#>    0.3441  -0.3146   0.8425   3.8142  -1.4193  -9.2191   4.1511  18.4759
#>    0.1444  -4.0375   2.6880  -1.7336   6.8913   1.2843  -5.5410   4.6154
#>   -1.3515   4.1558   8.1434   6.3881  -7.6126  17.2527   0.5874 -16.5160
#>   -3.3278   2.3915   0.1346   2.4815   3.1522  12.2366  10.4939  -0.2566
#>    3.0023  -8.0660   7.4737 -13.8353   9.4647  -4.1841  -2.6735  -8.8970
#>   -4.7870  -2.2907   5.9340   5.4414  -8.2643   7.4334 -11.9875  19.8308
#>    0.7898  -2.5121  -4.1297   2.4142  -3.0716   1.5231 -12.0679   0.1726
#>    1.3550  -1.3796  -4.6867  -5.0870   4.2036 -14.8771 -17.4144  -4.1443
#>    1.4266   0.4145   2.9968   0.6719  -0.9508  12.0718 -16.1224   1.2340
#>   -5.0006   6.2094  -4.3928   7.9415   0.4109  -6.7398  -6.2568   9.9236
#>   -1.2877  -3.0394   2.1409   4.1269   7.6279  -2.2672 -14.6187   5.4248
#>    3.6450  -2.6069  -3.1043   5.2003  -6.1129   4.0020  -0.4724  -7.6086
#>    0.4512   0.4182  -1.1923  -1.0733   3.7836  13.0975  -4.4361  -5.3742
#>    2.6630  -3.7710   1.6409  -2.9732  -2.9787  -6.8818   0.0845  12.6266
#>    4.0795   1.9635   1.4789 -13.2237  11.2880  -4.3489   2.4117  -4.2083
#>   -1.4521  -5.5397   2.3893  13.9924   6.5282  -6.6682 -14.9641   3.4581
#>   -0.4452  -1.6833  -3.0595  13.8559  -0.4321 -10.0805   9.0578  13.0909
#>    0.1034   3.3340  -0.4633   0.2078  -6.7895   8.3696  -3.6051  -2.0612
#>   -6.2375  -1.5755   1.9585   2.5183   4.4616  -3.5485  -6.4453  -4.5699
#>    0.1889  -5.0685  -0.3778  -4.1254  -6.4915  -6.3508   4.1894   1.0517
#>   -3.3687  -1.9437  17.7762  -0.2512 -26.5145  12.5644  -1.5246   3.2457
#>    4.5055  -6.5988   9.7927  -9.9268  18.2605   8.9813  -7.1450  -7.7198
#>   -1.7397  -0.9079   4.7866  -5.1672 -13.8461   9.3475  -3.6414   1.9066
#>    1.6295   0.0606  -6.5621   0.5167   0.4780  -8.8857   5.3690  -0.8721
#>   -2.1720   0.3425  -1.7462  -7.4614  -7.4939  -1.8299  17.2978   2.8358
#>   -1.1714  -0.7382  -5.4276  -2.4168   0.2923   3.3142  -5.7999  -2.3071
#>   -1.4656  -1.4142   6.6899   5.5320  -9.1698  -0.3805  -6.6655   5.8672
#>   -2.0749   4.6013   9.0970  -3.6577 -16.9117   5.5562  18.5035 -14.3026
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]