Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  7.8333e-01  4.7347e+00  8.9532e+00 -1.5195e+01  1.4039e+01  7.5323e+00
#>  -2.6210e+00 -1.6622e+00  9.7662e+00 -1.9803e+01  6.6495e+00  4.1747e+00
#>  -1.6902e+00  1.9897e+00  1.2169e+01 -1.3646e+01  5.8252e+00 -4.2507e+00
#>  -4.1398e+00 -6.7121e+00  4.4201e+00  2.0966e-01 -7.8694e+00  1.2754e+01
#>   1.2975e+00 -3.5774e+00 -8.2646e+00 -5.3024e+00 -6.5610e+00 -1.2453e+01
#>   1.3550e+00  6.9000e+00  1.0237e-01 -6.8567e+00 -7.2513e+00  1.1215e+01
#>  -1.1938e+00 -2.6477e+00  8.7162e+00 -4.4517e+00 -1.6851e+01  1.1233e+00
#>   4.9220e+00 -1.8981e+00  1.2232e+00  1.2198e+00 -9.5870e-01 -1.7485e+01
#>   1.0940e+00  2.7902e-02 -5.7712e+00  1.3011e-01  7.7783e+00 -1.2836e+01
#>   7.7698e+00  1.2801e+00 -1.0841e+01  3.2931e+00 -3.4824e+00 -1.2215e+01
#>   3.7867e+00 -1.6438e+00  1.5070e+00  3.8158e-01  3.6334e+00 -1.6923e+01
#>   3.4047e+00 -3.3918e+00 -4.6869e+00  6.7824e+00 -3.5374e+00  5.0465e-02
#>  -4.9754e+00  3.7684e+00  1.6034e+01 -5.2915e+00  8.0164e+00  1.2267e+01
#>   3.7365e+00 -4.3531e+00  1.1099e+01 -6.2208e+00  8.6898e+00 -6.1680e+00
#>  -4.0713e+00 -2.6317e+00  3.9943e+00  5.3872e+00  1.1087e+01  1.5248e+01
#>   1.1743e+00 -3.5589e+00 -9.1126e+00  9.9487e+00  2.8405e+00 -1.2505e+01
#>   1.9251e+00 -2.2543e+00 -3.8691e+00 -5.5056e+00 -2.1957e+00  9.1473e+00
#>  -4.7961e+00  9.3399e+00 -5.0443e+00 -1.6295e+00  5.7693e+00 -3.4493e+00
#>   7.4446e+00  1.8880e+00 -1.0178e+01  2.3580e-03  1.2051e+00 -3.5768e+00
#>   3.7021e-01  3.2081e+00 -6.7746e+00 -3.0880e+00 -2.0589e+00 -3.0689e+00
#>  -4.7100e+00 -4.8936e+00  3.2735e+00  2.6014e-01 -1.3387e+01  8.9493e+00
#>  -8.9189e-01  6.1282e-01  4.9648e+00  4.3534e+00 -1.5623e+00 -1.4183e+01
#>  -5.0809e+00  1.0463e+01  9.9972e+00 -1.3367e+01 -1.2809e+01  6.3438e+00
#>   1.2709e+00 -6.9952e-02 -3.4140e+00  6.0106e+00  5.6527e+00 -7.7861e+00
#>  -5.8817e-01 -2.4613e+00 -5.4372e+00  7.5578e+00  5.0551e+00 -2.2720e+00
#>   8.6049e+00 -1.4894e+00  1.6675e+00  2.9624e-01 -1.1439e+00  2.8991e+00
#>   6.1403e+00 -3.4797e-02 -1.2360e+01  9.6324e+00  1.3684e+01  1.1397e+01
#>  -1.1459e+00  2.0526e+00 -1.1838e+00 -1.3719e+00  9.1001e+00 -1.0883e+01
#>  -6.5965e+00  1.1290e+00 -3.0350e+00  9.7678e+00  6.6759e+00 -1.0245e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]