Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.9649   2.6441  -3.4294   1.0340   2.5842  15.6911   9.9366  13.7981
#>    8.1406  -9.4798  -0.5907   0.2777  12.3905  12.1650   0.3029  -5.4634
#>    2.0339  -4.3327   2.9882   0.0800  -0.7242   4.1394   0.9309 -12.1272
#>    0.0246  -1.9130  -9.7400   9.6930   1.6164   9.7363  10.8290   5.7312
#>    1.9618  -4.9565   2.7254   0.5433  -6.4493  10.2379  12.1690  -9.0156
#>    3.0331   2.1299  -5.7818 -19.4415   3.7107  -0.6353   4.8965  14.1859
#>   -1.3897   3.1254  -7.7146   8.2976  -3.7654   9.8793   9.1499  -2.1758
#>    7.5702  -2.3588  -3.9564 -11.2103   2.0773  -1.3015  -2.5460  -0.7128
#>   -2.3317   2.9378  -8.0219  -1.4114   6.0563   9.9623   8.4115  13.6591
#>   -1.3217   3.8342   2.2470  -5.6502  -7.6546  -1.4823 -11.0069   5.9206
#>    3.5613  -3.0931  -1.2092  -6.9986   2.3543   4.5531   4.5162   4.0857
#>   -2.1504   3.0362  -0.1380   1.0146  16.1250   2.7879  -6.0047 -11.1252
#>    2.1984  -1.1562  -4.2648 -10.6649  -6.9932  12.0977   1.1737   8.4614
#>   -0.3533   7.1319   3.6785  -5.1640   8.9019  -5.9922 -10.1683  -9.0828
#>    2.3367  -6.8252  -1.6931  12.4766   3.4809  -4.8929  -7.7727 -16.4282
#>    4.8136  -3.1833  -1.8216  -9.7724  -6.4296 -18.9493  -5.2045   7.6639
#>   -1.3342   3.4593   1.7013   2.2370   1.2380  -4.7057   1.3733   6.7359
#>    5.0031   0.2492  -9.9524  -3.9105   6.6645 -10.4727   7.4115  -1.5582
#>   -0.3169  -4.0841   7.5530   0.1234  12.2290  -9.0608 -10.7802   0.8842
#>   -8.5473   9.9594  -5.5344  17.6362   7.0710   9.2769 -13.3962  -7.1981
#>    0.9399  -3.0681   1.0492   9.4411 -10.2740   3.2762   8.1731   6.3529
#>    0.0119   4.3967 -10.1602   2.9235  16.2391   0.3607  -2.5361  -1.3715
#>   -0.1109  -6.0145  -6.8032   2.9472  10.9354   3.3373  -9.2976 -11.9751
#>    3.2027  -3.6939  -3.3546 -11.0303  10.0083  -0.9345  -0.8227  -1.2973
#>   -5.8083   1.9915   8.2123  -0.8691   4.4335  -5.6378   5.4427  -6.3076
#>    0.9330  -0.9253  -6.2112  -2.9840   1.8292  -2.3515   4.5458   0.2167
#>   -2.2875  -0.4734   1.6965   6.4813   9.2233 -28.7703 -17.7738  -5.6001
#>   -4.9361   2.2742   8.2119  -0.8486   5.6994   7.0082  -2.0643 -13.3474
#>    3.8746  -5.2185  -5.8954   8.8950  -3.2049   4.7877   3.5548   1.5502
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]