Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.2982   4.3867  -2.3836  -2.8135  -0.4230  -2.6163   5.4928  11.3869
#>    2.3282  -0.6908  -4.3527  -3.6899  -2.6447  -3.3143   7.3311   8.4258
#>    0.7281   7.4857   8.4574  -4.4296  -1.1587 -11.6168  -4.4813  16.2243
#>    3.5134  -6.2223   1.4294   0.9084 -13.8173  15.8139  -1.1920  -1.9373
#>   -0.0652  -5.8350  -5.4876  -6.8842  -4.1088  -3.0555  -1.8767  -3.4589
#>    5.0388  -3.8985  -3.9648  -8.3749 -15.8255   2.3670   3.5679  -4.4144
#>   -5.1950   2.5485   7.1966 -11.6797   2.2113  -6.0910  10.8517  12.0756
#>    0.2118   7.4496   1.5331  -1.4984   2.0444   8.6594   3.5240  -5.3373
#>   -1.6071  -8.8109   5.5555  10.9866   8.4914  -3.0764   2.7537  20.7022
#>    5.0308  -2.4130  -1.6303   1.4960  -8.1360  -3.9415  -1.6311 -10.5851
#>   -0.2228  -6.0431   9.3436   7.1050   9.0841   2.7509  -4.6193   6.5458
#>    0.5846  -1.8439 -10.0398  -6.1113  -1.4518   7.8384   8.4031   6.1635
#>    0.7494  12.5911  -1.7146  -4.6631  -4.3434 -14.9368   6.4755   0.1589
#>   -3.1442  -1.4589  -0.8348   8.7079   7.9083  10.2484 -12.3690  -2.1792
#>   -0.9649  -6.0300  -9.1112  -3.8204   0.1106  11.8748   4.1360  -5.3076
#>   -3.2551  -6.0159  -0.4095  -6.9795   1.1755  -0.4844  -9.6922  -8.0444
#>    0.0291  -0.7096  -5.1096  -9.3090  -0.4814  -0.3736 -10.9585   7.8569
#>   -1.9800   5.8207   6.6866   0.3350  17.6059  -5.8046  -4.0708  -5.2850
#>   -6.9086   6.7860   1.2489   1.2247  -2.1513 -26.3483   4.5375   0.7290
#>   -1.9727   4.5739  11.1259   0.5878  13.9559  -2.9719  -5.8484  -6.8689
#>    2.1097  -5.1020   1.9038   9.8939   0.7097   0.8959  -3.5715 -12.3226
#>   -1.8779  -1.3801  -2.3415  -2.6376   4.5366   5.1982   3.8207   3.3171
#>   -4.4774   1.1683  11.3951   0.5598  13.9260  -0.8029 -10.9959   5.0217
#>   -3.9951  -1.9571   9.9484   0.7439   4.6721  15.6417  -3.7317   5.9215
#>    9.2164  -4.4655  -0.0830  -4.2327 -17.1007 -11.1460   2.0479   6.6613
#>    5.6216   8.3603  -4.7466   2.8971  -8.2703  -2.7959  12.2642   4.7686
#>    1.7934  -0.9281   3.6840  -1.5035 -22.7815   1.8836   4.2808  -6.9812
#>    4.5644   3.7431 -11.0747 -13.5324  -2.7501   3.7345  -5.9629  -8.6348
#>    1.0071   4.7873  10.5617  -7.1387   3.9088  -8.0431  14.7989   8.1624
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]