Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.9203   0.9447   3.1182   4.2364 -11.6852  -2.6148   2.6116  -2.9052
#>   -6.9292   5.3589   1.5260   8.4678 -10.1504  -6.3158  11.8375   2.1655
#>    5.1215  -8.3741   1.5223   8.9126   4.0420   6.2930  -5.4941 -10.8952
#>    2.6400   5.7167  -0.2368 -11.4679  -2.5068   5.1463  10.8295  -8.1390
#>   -3.8924   2.1091   6.0908  -4.2725 -11.5551  -6.6607  11.4719  13.8864
#>   -0.5429   0.9636  -6.9439  -1.5238   2.1889  -8.1776   4.5243 -14.9717
#>    2.6273   3.1618  -2.2310   4.9236  -3.6775   0.8082  -6.7691  -1.8990
#>   -0.9571  -4.9149  -0.5321  -4.8259  -0.6903   4.9772  -6.3831   8.2615
#>    1.2317  -4.6304  -1.9288   4.4643   1.7542  16.2229  -3.8175  32.2796
#>    0.2758  -5.2853  -2.4276  -5.3419   9.7334 -10.5848  -5.3506 -13.7605
#>   -2.0011   2.9665  -2.6829   9.7362  -4.8618   1.8738  -1.0437  12.4163
#>   -2.7650   6.0013   5.7851   1.1840 -13.1675   4.0353   4.3354  13.8828
#>   -3.6317  -0.5637   2.5731  10.5512  -4.2235 -12.4257  -6.7225  -1.3671
#>   -6.1102  -0.9882  -4.6617  11.9687  12.1955   4.7911  -0.5029   9.4090
#>    0.1432   0.9209  -3.4532  -5.6301  -1.0925  -4.6267  -9.2198  15.2788
#>    4.2959   0.8295  -7.3659   7.4094  20.0654  -1.9514  -1.1335 -16.0386
#>   -1.3619   0.5512   8.9524  -2.8751 -10.4343   1.6473   8.1381  11.5482
#>    3.7669  -3.2620  -2.6944  -4.9087  -8.2381  10.0975   2.0247   2.2908
#>    1.7036  -7.9687  11.0233  -1.4277  -0.4775  -4.8440  -0.7004  17.6835
#>    1.2692   2.3969  -2.5741   9.0325   2.8580  15.3968   3.4177  -7.9598
#>    0.4968   1.0091   0.9005   5.4590  10.8830  -2.4140  -6.0294  -0.0081
#>    2.2290   0.0932  -6.5057  -9.5468   8.4640  12.2084   6.1667 -12.3056
#>    3.9804   2.7325   1.1083  12.7166 -13.6980  -0.9961 -11.0822 -11.0875
#>   -1.6992   6.7292   0.0416   6.0845   0.3718   7.3727   8.5089  13.3514
#>    1.1845  -4.6821  -7.6590  -2.9602  -3.3261  -2.0213  -9.8495  11.3793
#>   -1.2701   9.0769   2.4914   2.0126  -0.1255   9.6443   4.5355   4.8941
#>    1.9034  -4.5734   6.6191   7.1827  -7.1608 -11.2686  -9.2711  -7.9684
#>   -2.3677   2.4776   0.4119   0.8588   1.1624   7.2179  15.3016  -9.5002
#>   -2.7144  -1.3645  -3.0588  10.3677   7.5430   3.1749  -4.5494  20.9217
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]