Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.8128   5.9748  10.9499   5.4159  -3.3521  -1.7910  13.2450   5.8914
#>   -1.7378  -3.0011 -11.5473  -7.6660   6.4986  16.8724  10.1298  11.6807
#>   -2.2163  -1.7514  -7.9276  -7.2467  -2.9223   0.7218  16.1392   2.5599
#>  -11.6640  -8.6504  -1.9397   9.4744  -6.1510   1.7374   1.4021  -1.5546
#>    8.4121   7.0979  -3.1873  -6.9229  -5.6601  -2.4965   3.9616   5.5814
#>    6.1285   7.1732   9.7655  -3.6777  -6.7273  -4.1692  -1.8284   9.7826
#>   -9.1449   0.8428  -1.6043 -10.4326   2.2875   0.9479 -10.1378 -10.0831
#>   -6.2148   6.2818 -17.2678  -5.0967   2.1024   0.7543   8.1620   6.1292
#>   -0.7825 -15.4350   7.4427   7.4449   0.0384  -6.5091 -16.4401   0.5432
#>    3.2279   4.9345   9.0394 -10.7575  -9.1410   0.5377  -1.7454   4.0798
#>    4.2015  -2.2492  -7.1368  -3.8518  -4.1875   6.8565  -6.2296  -1.2355
#>   -5.6649 -15.0594  -2.8134  -2.2971   6.5899  14.7013 -12.0000  -7.9581
#>    0.4428   8.9745   6.0235   3.8249  -4.2781 -13.4364  20.2739  -0.2065
#>   -4.3066  -9.3178  -0.5807   5.3481   5.1044 -21.0580   4.2828   2.9743
#>    0.3858   3.3367  -3.8324  -6.9375   5.8824  -7.7493   0.8478  -4.2757
#>   -2.9999  -1.2411  -3.9532  -6.7766  12.8555   5.0311 -20.6548  -6.7809
#>   -3.9553  -3.7960  -2.9223  -1.3128  -0.5243 -14.4655   3.5232  -7.1109
#>   -3.8469  -7.5633   0.8366   0.9066   9.2217  -3.3501   3.9831  -5.3523
#>    6.6533  -0.5918  -7.4317   4.4052  16.0470   1.7232   1.5690   6.3107
#>   -8.6508  -3.7481  -2.2314  -5.4660   0.7577   1.5972  -8.6876   5.2105
#>    0.7131  -2.1481  -2.8321  -2.5520   0.2065   0.6258  -0.0197  -9.4216
#>   -0.3351  -1.3293   3.1907  -0.1647  -5.5314   2.2799   0.0019  -4.2354
#>   -3.6507  -0.8798   9.5639   2.4362   2.1756  -5.0401  -4.3455   8.1890
#>    5.7308   0.0286  -6.1843   8.6697  -8.4513   9.2727   2.9913  -5.9601
#>   -1.4127  -5.3470  -5.8200  -6.6347   4.4711 -17.8073  23.6745 -10.3388
#>    2.9685  -7.6891   1.4263   1.9149  -8.9515  -1.8605   0.0222  -9.3664
#>    4.8845  14.0165   6.0083 -11.8289 -11.4305  11.2388  13.1385   9.4891
#>   -0.4930   8.6638 -17.0566 -18.1697  -3.6199   3.7756   0.2442  -1.5894
#>   -4.3310  -8.6162   6.2546  -2.8880  -2.7969  17.2900  -0.3327   5.7703
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]