Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.5276   5.4413  -5.3899   8.9060  -2.6399   2.8136   8.9718  -0.1286
#>    1.4662  -0.4001  -7.5779   5.6730  -3.7354   4.8548  -3.2474 -15.7033
#>   -3.1896   3.6212   8.4503  -8.3361  -5.4869   0.0832  -5.3335  -0.1420
#>    5.7658   4.7672   2.2390  -7.8397  -4.6510   6.4966   7.0718   2.1332
#>   -3.9420  -5.4767  17.4488  13.0043  -7.4116   2.9530  -8.5746  -0.8281
#>   -4.0605   1.7285   6.4778 -16.0165   8.3855  -5.4923  10.4350  15.5516
#>   -5.0732  -1.3905   1.2567   1.0682 -10.6981  14.0020   1.1863  -2.1258
#>   -3.2330   5.3725  -3.7447   3.0426  -5.2488  -0.4121   4.6798 -11.8491
#>   -6.9645   3.7720  14.2038  -5.4168   7.8467   1.3278  -2.0843 -13.8035
#>    0.3189  -7.5826  -6.6740   4.1533  -2.7506   9.1899  -7.6400   6.0553
#>    3.5202  -9.6652  -1.2178  15.1044 -15.8343  -1.4269   5.1048   5.6261
#>    0.8971  -5.2267   1.5929   2.2590 -13.2539  10.4258  -2.0921 -14.1407
#>    0.5661   0.5490   6.0205   4.1358   3.9992   6.9748 -17.6600  -1.2835
#>   -0.9667  -5.0869  -5.9153  -0.3911   2.6581 -14.8613   0.7823 -10.1636
#>    2.8332  -7.3427  -4.4804  -4.2156  -2.7618   8.1420 -20.3164  -3.7851
#>    4.5929   6.9123   3.7862   0.8084  -5.6338  19.2382   3.4074 -10.0866
#>   -2.5461   2.0991   6.8052  13.9825  -9.0035  11.3019  14.5405  -7.6089
#>   -0.5363  -5.3664  -3.6036  -1.4296  -1.6948   9.8233   1.3753 -10.3050
#>    1.6295  -4.5529   5.6465   8.5958  16.5054   2.2311  -1.2033 -11.3432
#>   -1.1202  -4.3642   0.1571   2.4674   7.9527  -3.2314   9.8593 -11.4821
#>    5.2690   3.8339  10.6883  -3.0812  15.6223   0.1344   7.4472  -4.5718
#>   -1.2829  -5.5217  -9.2246  -8.7992  -3.2767  -9.3500  -3.7797  -8.6338
#>    1.0974   5.1602   3.3422   4.0217  11.5616  -4.0304 -10.5673  -2.3715
#>   -4.1791  -4.3054  -3.9663  12.4674  13.1558  10.8412  -7.0176  -6.3434
#>    3.5923   2.2059  -4.4517 -18.4074  15.0527  11.6530   0.8472 -11.0008
#>    2.3105   0.7091  -0.2127   9.2794 -13.9473   6.9834   3.0757  -6.4919
#>    4.9019  -8.4745  -8.7621   3.0339   4.7508   8.2534  -3.5957   3.9712
#>    0.8647   5.8695   7.7465  -0.4082  -4.3301  -2.1955   9.2980  11.2072
#>   -1.4905   4.5191  -9.7522  15.9950  10.9020  -4.3187  13.3527 -20.7109
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]