Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.5252   2.6460  -0.2172   3.5153  -6.8927   4.0253  -7.8073  10.7252
#>   -1.5380  -1.4596  -4.0526   4.9648  -3.3071 -13.3044  19.9016 -14.0183
#>    0.3810  -0.7534   0.6234   4.3556  -5.7950   0.6059   0.2897   9.5840
#>    1.8582  -0.4266  -7.7594   1.8790   7.7381 -13.4407  -5.4458 -12.9243
#>   -1.1924  10.7271  -5.0248   5.3406  -3.0378   7.9185  -2.9082 -14.9354
#>    2.8998   6.9497  -2.9181  -7.3057 -11.3938   7.6573  -3.2534   1.9424
#>    2.5485   5.8508 -15.5700  -1.9962  -1.6868  -0.3533  -5.6546   3.1745
#>   -1.2089  -5.8237   9.5108  -6.0000 -11.1823  -7.7044  -8.7669 -10.1434
#>    0.0958  -0.0213  -2.3430  -5.3578 -14.2429   0.9872  -8.2066 -10.1716
#>    2.1398  -5.9356  -7.3975   0.2044   8.8739  -8.0871   6.8471  -2.1905
#>    5.8144   4.9505  13.5629  -5.1140  -7.0307  15.0300  -7.5310   0.4716
#>    1.3036   7.9764  14.0153  -0.2349   8.5257   9.8719  15.4432   6.3271
#>   -1.8192  -0.0271   5.5117   5.4542   5.1620   5.0894 -15.7681   5.4832
#>    0.7113   4.9245  -6.2250   2.7749 -18.4373 -10.6724  -7.2132   1.2299
#>   -1.0192   4.3367  -2.9753  -5.2369  14.1495  -0.8524   8.3104  -8.7764
#>   -3.1533   6.3168  -3.2750  -2.8591  -3.3348  -5.3067  -4.8794  -8.3214
#>   -4.5166  -2.5954  -1.6334  12.4761   3.2956  -1.8328  -5.7165  -5.0710
#>    1.7524   0.0362  -6.2538  23.4095  -0.3505  -1.2826   7.8026   0.9744
#>    3.9576   1.0003  -2.3066  -4.4802   0.6276   9.6713   3.1066  -2.1462
#>   -1.0404   3.6475   5.7811  -2.4545  10.2797  -5.1224   1.4411   0.6771
#>   -3.0441   4.0594 -14.2846   0.7557  -6.8912   2.5273  -6.4523  -2.0275
#>   -1.5267   0.8087   5.7145  -6.1070   4.5120  20.2278  -3.9607  -2.4603
#>   -2.1648   0.0275   8.3036  -6.3800  14.6579  -5.6927 -10.0944   8.6476
#>    4.2631  -6.2496  -1.7600 -11.0741  -5.0704   0.7400  -0.0221  -9.7914
#>   -5.0011   3.0176   0.5827 -14.2732   4.1087 -15.5696  -0.2978  -7.4071
#>    0.2214 -11.6482   0.5125 -11.3542  -8.0601   1.4659   2.0082   1.3901
#>    2.9825   0.0041   7.2129  -3.9338  -0.7320 -11.2017   0.1502  -8.4434
#>  -11.4303   5.6856   6.9659  -5.1785   4.4536  -2.1598  -9.2737  -0.8010
#>   -2.2368  -2.3775  -3.8617  -5.0942  -9.5588 -10.4136  -5.5854  -5.7679
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]