Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.4098  11.1583   1.9565   2.0142   4.2317  -6.2156   4.3726 -13.8011
#>   -0.2139  -7.3119  13.5783   2.0125 -10.9938  -5.7269  -5.5574  -3.0185
#>    6.0197 -14.4030   3.5247 -14.8997   4.7178   6.1064  -3.9396   6.6772
#>   -6.9395 -10.4499  13.9861   7.7613   0.4724 -15.2077  10.4978  -1.0280
#>   -3.1785   0.5835  -2.4804   6.5895   2.8075 -10.5811   6.3787  -0.9789
#>   -5.0223   7.3931   6.9585  11.4143  -4.3907   5.1802   2.2582   5.9281
#>   -2.9600  -4.8083  -1.3706 -12.7476   7.0725  -6.8533  -4.5806  14.2742
#>    6.1139   7.3565 -10.2031  -5.6201  -1.1236  -1.9211   3.8096  -3.4299
#>    1.2250   0.2067  -4.3270   1.8699  -9.6754   2.5081 -22.7818   7.4724
#>   -2.7196   0.3157  -1.7577  -3.6714  -8.8921  14.4818  -7.3382  -0.9048
#>    2.3671   7.7165   4.7514  -1.0471   9.9884   9.5875  -7.4841   1.9428
#>    2.6961   4.8689 -11.8165  10.8229 -10.8896   1.6273  -7.2058  -3.2819
#>   -0.9007  -5.8807   1.4301   9.1670   1.8452  -6.2099   7.2173  -4.9021
#>    4.6163   0.4929  -2.1948   8.5314 -11.0056   4.9742  -4.0386  -6.3729
#>   -2.6130  11.4927   4.1919   1.2982   1.6135 -10.6011   4.7252   5.6473
#>    3.0665   3.7117   5.8570  -2.0720   7.8602   0.9864  -5.6196   0.3049
#>    0.4996   1.3074  -9.4862   9.8071   5.4746   7.1312 -14.9968  -0.6663
#>    4.4592   2.7545   5.0683   2.0967  13.6725 -11.9133  -2.6104  -0.3615
#>    0.3842   6.5638   2.0304   9.4468   6.3913 -16.6386  -5.9253  -2.7626
#>   -2.7608   2.5411  -2.5582  -4.6188 -15.0892 -11.9872   4.6438  -4.3259
#>   -4.5620  -4.8430  -2.3633  -6.8436  -7.9675  -6.0812 -13.5711  -1.8653
#>    3.9532  -5.9451  -4.1114   1.5860  -0.4572   5.7556  -6.5717   9.8730
#>   -1.5018  -0.3385  10.5251  12.0949  14.1130  12.6067  12.5933  14.4656
#>    4.7480   1.5105   7.3353  12.2345  -9.0035  -4.8272  -5.1591   8.0198
#>    1.5972  -1.2694   5.7517   2.0043 -17.8000   9.0896  -5.7273   4.2629
#>   -6.8750   4.0472   6.6210   7.0258  -7.2666  -1.7373  13.3651  -6.9037
#>   -1.3076   3.5686   2.4797 -11.7670  -2.8701  -1.1833   2.9981  -5.0331
#>    0.8305  -7.2408   5.0223   0.9448  -0.7206  11.2680  -2.5937   5.1837
#>    5.1149  -8.9391   5.6374  -4.9000   5.3801   1.2453   3.4430  -3.5323
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]