Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.0323  -0.6076  -2.0618  -1.5734   4.7115   1.6203  -8.9750   8.9802
#>   -8.7209   6.0764   2.3158   6.0047  -9.8460  -5.3807   5.9110   6.7313
#>   -6.7014  -0.3054  -9.8290 -13.4359  -4.5838  -7.3983  -0.4239   1.9748
#>   -0.5165 -11.1545   3.4774  -3.2832   5.6159  -0.1820  -4.6836  -3.1703
#>    1.4205  -1.6678  -8.8247   9.3677  -5.9747  11.2711  -8.3711   3.3824
#>    6.2900  -0.0224  -4.6156 -14.4919  -1.1415 -15.6129  -9.3947   2.3867
#>    3.5734  -1.0200   2.2472   8.6978  18.8428 -19.2310  13.3556  12.9765
#>    9.7455  -4.4216   3.8256  -2.8824  -6.3553  -3.5055  -9.6602  13.2835
#>    1.7950   0.1034  19.4872   9.2210   9.7755   3.0494  -2.7129   2.2143
#>   -7.5513  -4.8482  11.8698  -0.5401  15.7457 -13.6407  -2.4218   3.3975
#>    2.8736   5.0694  -8.8686  11.2064  -4.8760  -7.5669  16.7207 -17.9225
#>   -1.8904  -3.2459   7.9769  10.9622  24.6179   6.1685 -11.6198  -2.6627
#>   -4.2593  17.4386  11.1424  14.3541   0.3477   7.6412   1.6557  -7.8159
#>    0.6673   9.5016  -1.8783   6.6753   2.0248  -8.2228  -0.4495   7.6999
#>    4.2232   0.4700   0.6164   6.5306  -4.5141  -2.8423  -4.8510  12.7205
#>  -10.1479  -0.8851   6.0697  -0.8116   0.4589   1.9856   2.4825  -4.6863
#>    4.7818  -2.8435   0.9018   9.0365   3.8765 -16.4858   9.7937  10.6736
#>    3.5633  -9.5321   9.9328   1.5935   6.2687   6.8981  13.8736  -0.7782
#>    0.1360   7.9838   3.0917  20.6987  -6.1489  -4.1098  11.3333 -13.8567
#>    5.1587   6.8450  -7.7908  23.0038  -4.8895   5.5022  -3.9591 -14.0914
#>    0.2997   2.1466   9.6114  11.0398  -0.1943  -2.0650   7.4246  -1.0256
#>   -0.4422  19.0961 -12.6355  -4.6507  -0.3586 -17.9518   2.8579  -1.0132
#>    3.3288  -2.8429  10.4661  24.4494 -16.4787   1.4874  -8.6654 -11.5386
#>    7.3512  -5.7313 -13.2326  -7.3188 -18.3213   5.9084 -12.0932  -0.4034
#>   -1.0567   0.8249   3.1791  -6.5601  -2.9763  -6.9974 -19.9940  13.5608
#>    2.9456   2.5677   3.0343   5.6111  -3.9564   7.7263   5.6611  -8.3377
#>   -0.2634  -0.1560  11.8984  -4.5776  -4.0258   6.3654   6.0799  -9.1742
#>   -5.4020   0.1617  -3.4285 -11.4816  11.0287  -2.5087  -0.5997   8.3867
#>    8.2096  13.6082   4.2906  17.6097  -1.3696   0.6380  -9.3336  16.3386
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]