Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.5906  -0.2428  -8.8703  -1.0229   5.5181  -6.6682   2.4428  -4.4975
#>    4.5663  -4.0375  -0.1303   6.8421   0.5972   4.7016   5.4402  -5.6414
#>    4.0371   3.5469  -8.4411   3.5346  -0.9980   3.6167   1.0621  20.4635
#>    1.0247  -8.1446  -7.8689   9.0729  13.7282  -8.3384  -3.8124   5.8285
#>   -5.3889  15.6584  19.8602 -10.4660  -1.5446  -0.8138 -16.4520  -1.9380
#>   -5.1809  -2.3120   3.8864   3.0145 -16.3010  -8.0989   3.8076   4.9049
#>   -1.2753  -4.3611   3.4515  13.8180   7.6597   2.7928  -9.2398   4.0078
#>   -0.8189  -7.3263   1.3007  11.9379  -8.4660   2.0075  22.8351 -16.5807
#>   -1.7674  -5.5837  11.6244   9.6843   5.8146  -5.2547  -4.8431  -0.6837
#>    7.3862  -1.4384   6.2610   9.0750   7.5672  -6.0504  -0.8991   1.5020
#>   -0.9397   2.5150   5.6265  -9.2066   5.0777  -8.2720   1.5117   5.7149
#>   -7.8033 -14.3581  -1.3702   7.6596  13.3534  -3.9260   0.3654   0.3449
#>    1.4677   2.8370   0.5576  -8.4671 -11.9971  -0.6007  -3.1994  -1.1529
#>   -1.7366  -2.4707   0.4434  -3.1941 -16.8501  -1.0010  -0.3333 -17.7299
#>   -7.7018  -0.5266  -9.6456   3.1064  -2.9850  -6.2029   7.4738  -5.2400
#>   -3.3390  -6.0458   3.3308   4.4206   1.5412  -1.1068  -0.0474   3.9878
#>    6.8630   8.4250  -0.9302  -2.5178  13.2932  -3.0278 -11.0746  -5.3596
#>    0.2030  -1.4788  -4.3101   2.9099   3.2948  -1.0869  -1.0037   8.4457
#>    1.3877  -1.2555  -6.6553  -0.9536   5.3325 -13.1063  -5.3434   9.5600
#>    4.0198  -2.1808   3.2719   6.6275  -9.8734  -0.8764  -4.5025   5.1423
#>    3.9743 -14.3778   7.1658  -8.6220  -2.4994  11.5071   2.0209   1.3085
#>    1.2711   2.9668  -6.7120  -2.1876 -10.1998  -2.3209  -3.8484   3.6085
#>   -2.8091   7.4025   4.7028   3.4132  -2.5416   2.5002  10.4131   2.5338
#>   -1.9580  -2.4265  -3.7259  -1.7436   3.5180   3.3465  15.6230  -3.1640
#>    1.6604  -9.9044 -12.3413   9.4106  20.4501 -11.9949 -17.2757   3.8968
#>   -1.9503  -1.2318  14.7415   3.4984  -4.3617  -0.2658   7.3584  -8.5743
#>    6.8018  -2.2570  -8.2894   6.9476  -8.5303 -16.0043   9.7388  19.0133
#>   -0.5799   5.2137   5.6516  -6.2342  -8.7446  -6.4678  -5.0499  -7.4843
#>   -4.9250   2.1696   5.7509   3.5620   0.0324  -5.1111   5.2976  -8.4026
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]