Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  16.2442   6.6936  -2.2905  -2.5411  -6.6182  15.4901   9.3542  10.0713
#>    2.3698  12.8168   7.8922  11.7562  -2.5900   4.9432 -18.9629  12.4068
#>   -4.4598  -4.7496   4.3007   9.2941   2.6680   3.9349 -15.3332  -0.4550
#>   -6.2198   3.0003   8.7019   6.9423  -9.8945  -0.8851   3.3684 -10.5260
#>   -3.8359  -6.3477  -1.7850 -11.3905   1.9516  -7.7173  12.7408  16.4993
#>   -1.2094  10.7558   6.0852  -9.4390   4.9593   5.1948   2.9633  13.2205
#>   -2.0187   9.5185  10.3159 -17.6013  -1.7327   5.2291   8.7670 -18.1655
#>   -0.3249   3.3136  -2.9488  -5.3506 -25.0545  -4.4638  14.8194   6.5603
#>    1.1896  -2.3989   4.4316  -1.3453   2.5517   6.1253  -3.5927  -4.0829
#>   -5.1209  11.9651   8.7149  -0.1827  -4.2260  15.8812   4.6047  -7.1337
#>   -3.1709   6.6252   0.7047  -5.4829   1.0043   3.4155   3.8758   7.0241
#>   -0.9032   8.7440   8.2241   8.3985  -9.2179  -1.3726   5.5278  -6.9914
#>  -11.0184  -0.4880  10.8404   6.0736   2.6828  -1.0361 -11.2838  -1.1977
#>    1.1558   5.5777  -4.9482   5.3298 -11.2548   7.8380 -16.7908  20.6877
#>    3.9589  -5.6624  -1.7574  -1.2562  -3.1591  -6.9370  -4.0782   8.6504
#>   -6.7690  -0.8305  -0.1793  -6.1254  -1.0331  15.4718  21.8731   9.0069
#>    0.5664  -0.2388   5.3544   3.3593  -6.3708  -9.9774  -4.5092  -5.7556
#>   -0.5584   2.3845  21.5766  11.1109  -0.9884   1.9631   8.3805  -7.9575
#>    3.3573   2.8686  -1.0408   5.5656  -8.1912  -0.4088   5.6129   1.0154
#>   -7.0235   4.3450  -1.1156   4.3904  12.7463   1.9757  -4.9434  -1.6460
#>    3.7930  -9.3374   6.2127 -13.9527   1.5210   8.3385   2.5871  -4.6952
#>   -2.3990   5.6032   4.8448  15.3056   8.9393 -13.3311   8.3941   6.5633
#>    4.3268   0.0445   9.8830 -10.9849 -24.2736  11.1809  -6.7809  17.5278
#>   -1.9365  -8.1043  -6.0870  -7.2343   2.4571   6.9857  -8.4289  -6.5681
#>   -0.5647  -2.0679  -2.7483  -0.6904   7.1886  -0.7041   3.3234   2.6552
#>   -2.1714   3.7825  -1.3017   4.8864   3.7469  -8.2357 -10.9248  -2.8458
#>    3.9079  -4.3775  -8.3214 -27.3647   6.2671 -15.0353  -7.5877  -5.4262
#>   -3.7714 -10.4673   0.9544   9.6472   3.2177   6.8293  10.5333  12.3015
#>    1.7508   7.9287   5.7999   2.1091   1.9658   2.3681  -1.4093   6.3597
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]