Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -8.3051  -7.2525   9.4929   9.3181  -0.6930   1.5960  -7.2949  14.4239
#>   -4.5725  -7.4941  -4.0030  10.4430  -1.8154  -2.4903   2.5063   7.0009
#>   -1.7593  -4.8950   2.1175  -5.0303  -3.2646   9.2816 -19.1431   7.8074
#>   -0.7320  -3.0808  -1.5241  -6.3242   5.7494  11.5517   0.2007   4.7801
#>    3.6302  -1.6704   1.7437  -1.6437   2.8347  -2.0453 -11.9749   4.0694
#>    0.5273  -2.2056  -6.7844  -2.2462  -1.1977  -1.0178  18.7150  -7.6651
#>    2.9041   3.9622   7.1015  -3.1850  10.2409   0.7139  -4.5218  21.6770
#>   -0.7794   3.0257   0.9609  -7.6226  -7.2271   9.8444  30.6085  -4.6661
#>    5.0156   2.3568  -4.5260  -4.5083   0.3790  -1.5896 -10.6209   7.5920
#>   -5.9471  -7.5392  -4.1334  -0.7462   2.8284  -0.2216   2.5323  16.7114
#>   -1.1038   3.9798  -2.0873   6.3424  -3.5804 -11.3886  -2.9083   6.8531
#>    4.0598  -4.2230  -2.2570  14.9504   8.7007 -13.1996  -6.7171 -14.8900
#>    8.7770   6.4808  -2.9738   3.2675   2.8223  12.7754  11.4947  -2.4607
#>   -0.8850  -1.0787   5.7736   7.5695  -2.7671  -1.6151   7.6736   0.5578
#>    0.4751   2.1315   0.9150   2.6711   3.7515  -4.8039 -11.2405   7.8055
#>    0.4598   5.1494   6.1956  10.4621  14.1546  -5.3138   2.8272  -4.4333
#>    1.9487   1.0125  -4.1987   2.2684  -3.7275   0.1453  14.3631  -3.9208
#>    2.1198   6.0862   2.7473  -4.7883  -6.2328  11.6341  11.5564   6.4467
#>   -4.1684  -6.3366  -0.5922  -1.4869  10.4540  -6.6134 -12.0938  -8.7262
#>   -3.1938  -5.7235   5.4740  -8.1275   1.4589  -4.3775  -3.4327   1.2922
#>   -1.3183   0.3296  12.3441   6.0653   2.6619  -5.3450  -1.3595   5.4110
#>    0.2298  -8.8318   0.9737  -6.2272   4.6943   5.7085 -24.5347   0.6812
#>   -3.8219  -8.0560   2.1651   1.8468   6.8314   0.4286  -7.2915  18.5034
#>    0.3372  -0.2648  -1.5361  -3.6606  -8.5532  -4.4664 -13.2348   3.9704
#>    1.1167  -3.9570   4.7970  -1.8902   3.0745   4.4356  -7.2895   2.4185
#>    1.2636   1.8732   3.2296   1.6265  -4.8638 -15.0295  22.4238   2.6102
#>   -0.7039   1.4872   1.2960  -4.6977  -1.3951 -19.3960   6.6140   0.7290
#>    4.4388   0.1025  -2.6312   2.1651  -8.9034   9.1047  -3.9310  -5.9063
#>   -1.8824 -10.5324  -2.4239  -3.8827   6.7351  -7.7980  -0.3456 -12.9434
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]