Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.5315 -10.8733   8.4753   5.2567   0.2083  10.2961  13.6360  -3.2741
#>   -4.0714   1.6699  -3.6055  -8.9373  -0.7560  11.2783  10.9280 -11.1445
#>    7.3151   5.3695   2.2856   4.8652  -5.5095  -8.3948  -9.9833   4.7159
#>   -0.2524  -5.8583  -2.1860   3.1078 -11.2985  -7.0307  -8.9428  -9.3609
#>    4.6630  -8.4680   5.0260  -1.7422  -7.4651   5.6153  -5.0543  -8.5046
#>   -2.0606  -9.9409 -15.0010 -10.5788  -2.5567  -1.6100  -9.9098 -11.0905
#>    0.8054   4.9270   9.0163   5.7631   9.4303  -1.2747   8.2318  -2.8404
#>    1.5436  -4.3129   0.3834   9.8050   1.2881   0.2413   7.2073  -7.4766
#>   -2.4194   2.8002  11.7951   1.3724   1.7214  -6.1993  -3.0028   4.2823
#>    4.5750  -8.7564  -6.4449  -8.4929   3.4090   0.2172   6.3269   8.5196
#>    0.7493   8.0280  -0.9627   9.3477  -7.0556  -2.8010  -7.8634 -19.5773
#>   -3.1835  -0.8840   1.2974   2.9706   7.3600  11.6242   2.4362   6.3302
#>    2.0197   0.7622   3.5790 -10.2274  -9.4399  -2.4574  -9.1232   1.4016
#>   -1.5086  11.2323   3.3666   1.8406   0.3351 -14.6875  -8.6997   6.2694
#>   -8.2172 -12.7153  -5.9603 -11.7729   5.6449   4.9356  14.5965   2.3375
#>    0.5058   1.6956   5.9956 -12.8700  -3.0412   4.0198  -7.3765  -0.1938
#>   -1.0612   6.7622   2.0504   4.4456  13.5797  -4.1066   0.3987  -0.1754
#>    5.1566  -1.6954 -11.1187  -2.9657  -5.7608  -6.6912  16.6560  -2.8319
#>   -2.2090  13.2628   1.0877   4.2088   8.8017 -10.2733 -12.9827   8.5676
#>    0.7479   0.1082  -5.9589  -7.3305 -14.4883  -1.9069  15.4198   8.2245
#>   -3.2979  -0.2779  -3.6661   5.5725   6.9380  14.1214 -15.6750   4.5285
#>   -5.3992   1.4897   6.0098   5.8885   8.1552  -2.0237   0.2424   7.8136
#>   -6.9881  -8.5025  -1.4215   2.6460 -12.2855  -3.3126  14.5127   0.8305
#>   -1.0012   4.3430   0.1114  -2.4777  -7.2285  10.3742   3.2101 -20.4532
#>   -5.3060   1.2673  15.1424  15.4924   2.2583  -3.5412  -1.4822  14.7803
#>   -1.2322   4.0174   0.6064   5.1338   2.4964   0.1672  -6.4234   0.7899
#>    4.6428  14.9408  10.0840   6.6275   1.7833  -0.2886 -13.7369 -15.3607
#>    3.6028 -14.5382  10.7849   0.5787  -1.7901  -3.6510   5.5142 -11.5443
#>   -3.0179  -0.9807  -7.3504  -8.7887  15.7109  -4.7847  -8.6719   5.0956
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]