Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.6815  -6.2497   1.2750   0.8726  12.6583  -7.2666  -4.8364  -0.1835
#>   -2.1965 -10.8902   5.3108  -0.1930  -6.0981  -7.9410   3.2056  -3.7272
#>   -0.4429 -11.2885   1.7218  -3.9640 -11.0263   8.5908  12.3137 -25.6041
#>   -0.6467  -2.4994   6.5266  -0.4871  -8.1468   1.3607   6.4900  -1.1498
#>    5.4049   5.4127   4.5208   1.0333   3.9126 -14.4032   1.1140 -16.3417
#>   -2.7495  -6.4261   5.7618   5.4444   1.5134  -2.4813   7.2922  -1.8156
#>   -1.8552   2.2692   3.7490  -1.1165   0.6883 -11.1495  -4.1740 -10.0695
#>    2.6873   2.3028 -14.4529  -3.4004 -11.3061   5.6294  -0.4661  -3.3752
#>    1.8801  -0.1951  -0.1446   1.5434 -14.5804  -2.7924  -5.5415  -9.4745
#>    6.7454   2.7838 -18.4089  -1.0610   6.3717   8.6010  -6.4215  -6.1700
#>    2.2260  -3.5055  15.1746   5.1826 -21.5235   5.4127  -1.2733   6.2855
#>   -4.1529   8.7289  -1.4230  -7.4656 -10.4289   2.9100 -11.0689   4.9728
#>   -2.2882  -6.1675  -4.0326   3.3038  -0.2555  -4.5217   3.5972  -6.7810
#>   -4.0440   0.0477   9.3356 -11.2743   1.2250  -2.0105   5.1383   4.3459
#>    1.2191  -6.4578  -0.6448  -7.9792  24.6419  -2.0767  -0.7034  -4.8769
#>   -2.0676   1.9850   5.5209 -19.6479  -1.7856   2.0848  -1.5630   7.5710
#>   -0.8618   6.4848   6.0224   3.4875  -0.8948  10.2569   7.2094 -11.1838
#>    2.8348  -5.8019   5.1284  -1.0931   5.7336  -3.9071   2.2043 -11.1680
#>    4.3082 -10.2403  -1.5202   3.3533  -6.4495  -6.3943   5.6392   7.0673
#>    3.2854  -3.3014   7.5264  11.7073  -3.7606   3.2688  16.7240 -21.9912
#>    5.0605   0.2865  -6.5424   5.6327   5.3979  -7.2248  -4.1015   9.4810
#>    4.1304  -1.5301  -0.3066   3.8268   0.7063  -2.3997   6.1532   6.3088
#>    0.4471   6.7678   2.8871   0.3217   7.4855   3.2327 -16.0943  -4.4439
#>    3.2087  -5.2829   0.7671  14.8569   3.4189  -0.8218  -1.1038   5.0579
#>   -2.2998   3.9721 -14.2664  -2.8089  13.5883 -10.8669  -0.3010   8.3999
#>    2.4189  -4.8888  15.0306 -12.3650  10.6295   3.7035  -4.3833 -13.0256
#>   -3.0413  -4.5486  -2.9789  -3.7617  -4.3150   4.3443  -4.7451   7.2098
#>   -0.2625  -3.8694   7.5799   2.7042   7.3291 -10.1737  13.0341  -9.8871
#>    5.1258  -7.5019  -2.5054   7.2933 -13.0230  -9.4000  -3.1667  -9.9903
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]