Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.1286   4.7593  -0.2757  -4.3097   2.6355  -2.8696  10.8734  -1.3153
#>   -2.0820  -3.5223  -6.1399  16.9331  -1.3936 -20.0397  -2.7311   9.9222
#>    4.7777  -8.4682   3.6349   2.6749   5.1529   9.4179   9.8626  -1.5725
#>   -1.9078  -3.2520   5.0347  -2.0057 -16.6173   3.4244  14.1794 -11.9762
#>    0.4849  11.7783   4.3596  -0.7967   0.5901  -0.9702   2.1131  -7.8349
#>   -1.7738   2.7956   6.4345   0.4653   2.0620   9.3200  -3.6173 -12.3696
#>   -3.2331   1.5334  -0.2424   2.5087  -4.6857   7.2134  13.9017   9.1072
#>    2.5964  -4.1127   2.2952   9.4494   4.1100   1.6408  -2.9661  -3.2808
#>   -2.6051   0.1507  -1.2897  13.5861   9.5700   1.4568  -6.1839  17.0903
#>    1.8656  -0.7499  -3.0394   6.4827 -10.5005   9.5914  -0.7965  -0.6623
#>    2.4563   1.9464   1.9486 -15.6874   6.0486 -12.8719  10.9451 -11.6302
#>   -0.1105  -1.5081  -7.4614   3.1395 -22.1323  -0.1625  -2.8359  17.8829
#>   -1.5998   1.1943   3.4234 -10.6026 -11.0299  -7.0477  -0.2275  20.0819
#>   -3.0317  10.5021 -12.9929   1.7902  -9.7256   8.4912  18.4897  -7.4625
#>   -3.8939   9.4730 -15.5400   6.0235   5.0220   5.7224  -5.6285 -11.8303
#>    9.0784  -3.0184  -4.0995   8.1224   1.0602  -8.2913 -10.7623  -0.4076
#>    2.2364   4.6696   3.1384   4.8929  -7.8568   9.8071   0.4753  -4.3890
#>    3.0637  -2.0562   6.2124  -5.5236   7.0285  -0.7313   6.2070  -7.7430
#>   -0.4572  -8.2180  -4.0993   5.5373 -20.7942  -3.9707   9.9595   1.8169
#>    2.4852  -4.2081   4.9478 -16.2493   6.5611  -3.8756  -0.1643  -5.0656
#>   -0.9454  -5.3138   8.5414  -8.6642   3.6912  -1.4178  -8.0158  13.5988
#>    4.2801  -6.2200  15.7314   0.0585  -7.2230   1.2770  -0.2164  13.5159
#>   -1.3755   1.1713  10.9578  14.4037  -5.4408  28.1959  12.8041   3.5834
#>   -0.1331  -7.4315   1.4771  12.4449  -8.2092  21.2969 -10.3466  -6.8546
#>   -6.8004   6.2640 -13.3947   1.4882  14.4054  -3.6488  12.9805  -7.6274
#>    0.9304  -1.2883  -6.1720   4.3754  10.6681  -2.9125  12.6882  17.9737
#>   -2.5652 -11.8930  13.6476  -9.8477   0.0125  -2.3703   4.7368  10.8192
#>   -2.0088   4.9288   3.8394  -8.8465 -13.2215  -1.5435   1.7248  11.6943
#>   -2.4219  -2.2349   2.9548  -4.1301  -8.8049  -7.6280  10.9017   6.8049
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]