Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.5992   3.1860  -3.6961  -5.1334  -7.5617  -0.8757  -2.6984  -9.2827
#>   -1.3798   3.3849  -8.4912  -5.5011  -3.9146   4.0707   9.2838   7.5902
#>   -0.6259  -3.6433   8.4051   1.3395  -5.7253  -7.9426  -3.6679 -14.6170
#>    0.4173   4.3157  -0.2138   4.9802  -2.8564  -2.8285   5.8767  -5.0561
#>   -0.2410  -2.8321  -5.8461   8.1872  -7.4960  -3.0793  12.1579  -0.2067
#>   -4.2385  -6.1111   6.6828  11.4578  -2.4421  -8.8806  -7.3877   4.8395
#>   -0.0188  -3.6824   6.4930  -3.2646   2.5064  13.8670  10.1098   5.0934
#>    3.5402 -10.0892   2.7497  12.5495   5.2158   1.9485  -8.9009   0.8423
#>   -1.0061  -1.2348   4.9644  19.0753  -5.9777  11.9711  -4.8446  -0.7512
#>   -2.7534   1.2048  -2.8078  -1.1356  13.0457  -8.3615   7.4627  -3.6393
#>   -1.8114   0.0727  -1.4696  -2.7701 -12.7314 -18.6967 -10.3083  -5.6003
#>   -2.9249  -1.4174  -4.7054 -11.2107   0.6318  12.0303  -0.4820  -2.2858
#>   -0.8996   0.3920  10.6697  -5.7217  -2.7138   1.4282  -6.2279 -20.1820
#>    0.0148  -0.5094   7.3096   0.4580  -3.5934  -4.2262  -1.8763   3.6865
#>   -2.5110  -5.9534  -1.1763 -11.0615 -12.1725  -8.3702 -15.8334   3.0895
#>   -0.0952  -2.3767   0.5526  12.4229   2.6015   0.0623   0.8888   9.3520
#>    0.2528  -1.3111  -3.9485   0.3587   6.5337  -1.2944   9.5231   0.7427
#>    2.5928   1.6886   2.5648 -11.9700  21.3072   4.7759   0.2782  -3.2697
#>   -0.9331  -5.4017 -16.6626  -0.1558   0.0579   5.1015   2.4413  -6.4250
#>   -2.6352   1.7443  -2.0312   0.6026  -1.7483   0.2558  -5.7305  -4.0742
#>    0.9492  -3.5571  -5.9554 -12.0704   2.3860 -12.5838   4.7221   3.5658
#>   -4.2795  -3.3281  -6.3832 -12.5838 -11.0601  -5.4890   8.9876 -12.4539
#>   -0.4530   4.8842  -4.1149   3.6823  17.9984  -6.8053  -1.5900 -10.2357
#>    3.7302   1.0565   0.7607   2.5663  -9.1409  -2.5798  -8.8021   2.8584
#>    1.0998   9.6326  -0.9533  -0.3584   2.9691   6.6149   6.5576  -6.7291
#>   -0.4354   6.5732  10.6289   3.7271  -3.1693  -5.3092   3.8255   9.2219
#>    1.4908   5.2684   5.3202  -4.8906  13.3050  20.8163  -8.4275   4.4531
#>    4.0832  -2.7001  -4.4607   3.3560  -3.9414   3.9244   9.6407   8.6145
#>   -0.9957   3.6837   1.5536  -4.6838  -6.4808  -2.8927 -11.3096 -15.7452
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]