Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.1198   0.1892  -0.5193  10.1188  12.5451 -12.6223 -11.2144 -17.6908
#>   -0.1631   0.0195   3.6722   1.7795  13.4208  -0.2270   6.0589 -11.2461
#>   -2.5552   1.9130  -9.5401   2.1589 -15.6656  -0.4330  -4.3578  22.8457
#>   -4.3570  -7.2135  -7.9609  -3.6290  -8.2947   0.2210  -1.9235   2.5280
#>   -7.0382   5.7184  -6.3672  -2.8753  -6.7822  -2.5454   9.4666  -3.9271
#>    4.8213   0.9780   1.9746  18.2366   3.0048 -10.4352  10.6460  -7.8860
#>    1.1185   3.9610   0.5841  -4.6650   1.8332  -0.0072  -1.3484   4.7565
#>   -1.1799  -3.0091   6.4504   9.0365   6.5260  -8.0337   4.0205  -4.1667
#>    0.7211   4.4791   4.0060   4.7161  -1.2675   4.4997   2.1193  13.7634
#>   -4.8920  -4.7406   3.1681  -1.8089  -9.3894   9.2135   8.8442  -7.0560
#>    4.9521  -5.4237  -0.4431  -2.5385 -16.1535  -0.9031  14.6793   6.5968
#>    0.0799  -0.0665   0.1055 -15.1341 -14.6247   0.1101  -4.2773  13.2763
#>   -1.8550  -5.5507   7.8226  13.5421   8.0746  -5.9584  -4.4430  -7.1976
#>    0.6519   3.4137  -4.4433   2.7384 -11.1782   2.4530  -0.4462   1.0708
#>    0.6088  10.9473 -16.4069  -1.5419  16.5248  10.8229  13.0337   3.3501
#>   -1.0876  -1.5451  -8.5066  11.0528  -0.7171   9.1350 -13.3485  -4.8633
#>    4.0020  -3.1802 -11.0761 -13.9752  -1.8255   9.8655   4.5503   7.0352
#>   -0.0664  12.7215   2.7102  -4.5325  -5.6780  -5.6279   1.8663 -18.8872
#>    3.0767  -7.8869  -5.6365  11.5427  -3.9678  14.2981  10.1706  -5.2874
#>   -1.2322  -2.8361  -4.2734  -3.5504  15.1158   4.4697  19.2647 -10.2702
#>    0.8725   4.5584 -16.5133   6.6954  -2.9516  -8.6000   3.7962   8.6724
#>   -0.7399  -4.1565  -3.2834   3.5600   4.9878  10.4312 -11.0699 -14.7887
#>    3.6493  -4.2645 -10.0504   3.3008   7.7371  -2.3897  -1.2195  -6.6059
#>    2.3572   5.2088   0.0496   4.2755  -1.1017 -10.3591  11.9193   5.1479
#>   -3.7827  -5.1792   5.8772  -6.4698   7.9988   0.9716  -4.0204  -7.1624
#>    2.7610  -4.0340   3.2939   5.4449  -1.1299  -0.9151  19.0618   7.5429
#>   -4.9694   7.6935   1.0379   6.1653  11.6198  -0.8544  -4.7926   0.6049
#>   -1.4777  -1.4168 -10.3384  -7.7848   2.5830   8.7678  -2.7003  -2.1666
#>   -0.2174  -6.2545  -0.5624   9.7586  -3.3983   8.6259  -1.4678   2.4259
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]