Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.4777  -4.7903   7.2821   1.7539 -10.3426   5.6938 -12.0695   1.2350
#>    5.9389  -1.4484  -3.0537   0.4933  12.4195  15.1375   3.2271  16.4515
#>   -1.8188  -2.8250  -4.2961   8.0938   1.0044  -5.3851   7.6132  22.2417
#>   -0.7309  -2.8668  -2.1056  -5.7734  -5.4975   4.1135  -4.2855  10.6297
#>    4.7853  -9.3067  -1.1711 -10.1921 -10.8959   7.4540   6.2327  10.0514
#>   -1.8789  -2.9353  -0.9036  -1.3907  10.6974  12.9332  16.2653  14.7149
#>    3.7422   2.1565  -3.5017  -4.0253   9.0856  10.6365   1.9412  18.6276
#>    4.4057  -0.5024   1.3014 -16.0436  -4.2553  -8.9692 -13.7194  -3.8055
#>   -4.9652   9.0935   9.7437  12.1178  -1.1048  -8.7456  12.1153  -4.5620
#>   -1.7021   5.1096   4.2914   7.1726   6.1643   1.4273   7.6008  -1.2040
#>   -3.6233   2.2398   2.5027   8.6854 -11.5201  -0.0948  -1.3405  -8.6092
#>    4.1848  -7.9980   4.3797   2.2863   9.7466   3.5515  -4.2583   2.9881
#>   -2.0561  -4.7143  -0.3000   3.7803 -14.0671   3.6581 -10.0396   0.4173
#>    0.0019  -0.4164  -8.4011 -10.3847  -7.0663  10.0254 -12.2700   8.5565
#>    4.6616  -3.2934   1.7902   1.9110   7.8525  -9.2454  14.3396   6.5069
#>   -4.4296   1.1429  -6.1079 -18.7938  -2.3867  -5.8561 -11.1301 -12.7569
#>   -7.3894   5.5642  -1.1461  -3.4130  13.9603   5.5604   1.5899   9.9802
#>   -1.2611   2.4215   6.9851  -4.6288  -3.0743   6.9655  11.2963  -1.1338
#>    8.2270  -6.1048  -5.6683   5.3784   0.0654 -16.9294  -6.4864  18.7702
#>   -3.0209   3.4819  -4.6028  -4.2863  -7.9256  -0.4449  -4.4415 -11.1677
#>   -1.4781   1.2447   3.3562 -10.9083   5.3508   8.2688 -18.9740  13.7673
#>    4.0043  -8.6086  -6.3814 -10.5292   2.0668  -7.9229  -9.8110  -4.8387
#>    2.5971   0.2420  -1.2366  -3.1469  10.4149  13.5789  10.7140   4.5763
#>    2.5211  -6.8452  -0.0141   3.5405 -17.3902  -3.6655   4.5249 -15.2803
#>    0.0885 -11.8785   5.2118  -3.5427 -12.9497   1.1707   9.1657   4.4630
#>    2.8489   4.0015   0.4128  -7.1729  -4.7338 -10.6579   0.3028  -3.0446
#>   -0.9431   2.0565   8.2036  -0.6468   6.9929  -7.1574  -8.6598  -5.0761
#>    0.0186  -2.3571   5.0369   1.0709  -8.9651   5.2925   4.5862  -6.6852
#>   -0.2711   1.6166  -6.8845  -5.8297 -18.6338 -14.6070  11.2371  -7.8318
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]