Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.0716  -4.2481 -10.8251  -2.1075   1.5414  -6.0724  -4.6257  -4.5045
#>   -1.6799  -2.7323   8.1156  12.8748 -17.2490   3.8091  13.2557 -18.9928
#>    3.2462   8.2430   9.8245 -10.9752   6.3177   0.2030   0.5781   3.3538
#>    2.6673   1.5723   1.0723  -6.2365 -20.1893   3.1424  -9.0650   2.7054
#>    3.7649  -8.9377  15.9336   4.0136   0.6102  11.6663  -0.1092   9.8733
#>    1.2062  -0.1268  -5.8340  -9.5883  -4.6816  -4.0940   4.2887 -15.2182
#>    7.2613 -11.4602  -6.5500  -6.4122 -16.7727  -0.1035  -1.8101  -4.9905
#>   -2.3536   2.4932  -1.0654 -10.7033  -1.9435  -4.0391 -13.2001  -4.4506
#>    6.2112   2.7836   1.4507   4.3722  -6.8588  12.9431  -3.4223   0.9982
#>   -0.1897  -2.3334   8.2582 -12.6615  -5.2927  11.7999  -7.2343  -2.6190
#>   -1.2918   0.7694  -5.0076  -1.3905   5.8934  -2.9554   6.7140   4.4904
#>    2.5269  -3.9741   1.8233  -8.5582  -1.9430  -0.5783 -21.2008  -2.3847
#>   -3.9327  -7.0380  -4.1016  -9.4495   2.7044   1.4526  -4.1966  -1.0252
#>   -6.9327   4.2780  -0.1703   3.9435   1.4641   2.6742   5.3455  -9.4691
#>   -1.5421   9.4176   0.4836   3.5201  -7.8013 -11.4436  -6.6894  -4.5425
#>    0.5480  -3.2666  -8.7378   7.6247  -2.2515  -3.1919   4.6447 -13.8567
#>   -1.8903  -4.9294  11.5654  -4.0672 -16.4460   8.8100  -2.4037  -3.0729
#>   -0.6200   3.5271   9.3807  -1.7378  13.2339   4.2372   0.0479   8.8012
#>   -1.8174  -2.3512  -4.6417  -0.2968   7.2371  -6.2870  15.0526 -15.5018
#>    6.9052  -3.3316 -11.9059   2.2822  -3.9601  -3.6934  11.5364  -1.7938
#>    5.1776   1.9291  12.3053   2.0337   2.0093  -0.6980   4.6883 -11.0441
#>    0.0413   8.2613  -2.5490   1.6627 -10.9886   0.3932  -6.0379  -1.0246
#>    3.6375  -7.0594  -1.5835   1.1734  -6.8880  20.6799  -8.6644  -8.8446
#>    1.5799   1.1006  -8.6644  -2.4264  -1.0586 -13.8255  -4.9683  -1.0488
#>   -5.7042  -2.4773  -5.6584  -4.9055   0.2894  -1.3340   0.7514  -7.4509
#>   -8.3992  -3.4264  -1.2291   3.4404   5.4654   5.4102   3.1158  -4.1432
#>    0.6089   1.3494  -8.4020   7.7882  -1.1021 -10.8113   3.2934   5.5394
#>    7.8586   0.2991  -5.7214  18.6618   7.8839  12.7587  -3.8110   0.3816
#>    2.6765   2.7936  -2.7262  -3.9499   8.0721   4.4610  -2.6160   4.4838
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]