Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.3675   2.0439   2.0745  -0.5995   1.0857  -9.0139  -4.2181  18.4080
#>   -5.9855  -4.1662 -11.9214   4.6223  11.8534   3.4845  -6.8342  -6.3532
#>   -2.6479 -11.5420  -1.6416  -1.7705  -8.1872  15.7681   3.0400  -8.4711
#>   -0.7076   8.7019  -7.6910   6.3891   4.9061  -4.9722  -1.7164   8.2617
#>    3.0546  -9.1564  -7.4228   6.7211  -1.2477   2.9808   8.2626   2.4870
#>   -1.1309  -7.2083   9.0978  -8.7699  -8.5625  -3.4181   4.3485   4.0502
#>   -2.3425   0.0056  -1.4468  -2.7997  -5.4375 -13.4929   6.7544   0.7636
#>   -9.9237   3.7017   1.2465   8.6176  -4.3662 -12.7306  10.3304  16.8266
#>  -11.7016   1.6833   8.2937  -4.2836  -7.4391  -0.1085 -14.0002   9.7347
#>   -5.9119   4.6250 -12.1309  16.4564  -6.7953 -11.1305   2.3045   4.7925
#>    0.9804   0.0309  -6.8582   2.5881  -4.8579   1.7513 -12.4650 -12.0992
#>    4.5289  -8.5433   4.7676   2.3779  -5.8957  -5.3239  -0.4333   2.4298
#>    8.2446   6.3510  -3.9099   5.1449   2.8367  -2.7243   4.6406 -20.0126
#>    2.8946  -6.3494   6.8599 -11.5978   1.6257   7.5273  -0.4455  -5.7366
#>    7.3413   3.3572   7.5908  -4.8701   4.5370  16.3715   5.2916  12.8937
#>   -7.9974   3.4790  -5.4779  -2.5964   7.1132   0.7722  -2.1926  12.8515
#>   -0.7891  -6.4224  -4.0440   4.2722  -0.1162 -12.4960   8.5508 -15.9245
#>    3.7900 -11.3330  16.1657   3.9002  -8.9286  -2.3878  15.0470  -9.4464
#>   -0.5793   5.4105  -9.0626   2.5814   4.2293  -5.9990  -3.4820  -1.1525
#>   -3.2001   3.1107   4.7364  -8.0638   7.8332  -3.5923 -13.7074   4.7440
#>   -4.4888  -6.5558  -2.7779  -6.8382  -3.2404  12.5161  -5.2406   8.9394
#>    1.2640  -2.8358  -9.5791   5.2228   3.7309  -1.0974   5.0252   2.7319
#>    2.9975  -2.0199   0.9099   2.3146  -2.0225   2.3910   6.9382  -8.3718
#>    1.5872   0.3228  -8.6278   7.1820  -4.4297   6.3280   6.7009  12.0768
#>   10.4978  13.8467  -7.9136   4.8748   5.1596   2.0287  12.2489   1.9805
#>   -0.1243  -0.0592   2.1671   3.1965  -8.9449   1.8242   9.8037  -7.1351
#>    7.8331   3.6119  -5.8452  -1.3522  -2.7851   4.1730   9.1893   4.8539
#>   -5.8805   2.9654  -7.7535   3.6018  -8.1271  15.8651   2.1490   0.7800
#>    5.3678  -7.5701   0.7153  -6.9157   0.4957   1.8144  20.4915   2.5979
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]