Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.5216   1.7456  -0.9786   5.7296  -5.3081   2.8208 -14.3893  -1.9445
#>   -3.5887   8.9340  -2.9663  -3.5981   5.7889   4.3277  -1.0459   2.6172
#>   -1.9333   0.3703  -8.0824 -10.6848  -5.5777  16.3393  -1.5610   0.7399
#>    3.0636  -1.3565   9.1833   5.8960   4.8152  13.9044  -5.0560   4.5321
#>    3.9114  -8.4139  -4.7006   4.9873   1.3197  -1.3232  12.5776  -3.8000
#>    2.7036   3.1941  -3.7927  -6.2684  -3.0899  -7.7296  -2.9192  -1.5382
#>    2.8753  -2.5719   8.7048   6.4651  10.5408  -7.9084  -2.6364 -16.6518
#>    2.9737   2.4047   0.7732   0.7294   6.3330   8.9967   9.3986  18.7636
#>    3.1607  -7.0476   5.1084   5.4081  21.9564 -12.4035  -8.6797  16.1043
#>   -0.8557  -1.7580  17.0926 -10.6074 -14.0547  11.4495   3.8562   7.3023
#>   -5.4457  -2.7180  -2.2735   1.4464   4.8584   7.5052  18.1690   2.6649
#>    2.3182  -6.1082  -9.4072  -8.6923  -5.9859  -5.7856  -7.1745  -1.3905
#>   -0.7335   9.5796   4.2187 -15.3284   8.0439 -11.5754  -4.6379  -2.8874
#>   -4.7858   4.8307  14.2454  -7.2308   7.2496   4.9017  11.8708  -1.4750
#>    1.5061   5.3546   2.5775   0.4620   5.1780   7.1680   4.7944   2.8273
#>    3.2941   1.8884  -9.9540   6.4062   2.2225  14.8311  14.4985   9.5560
#>   -2.4868   4.3939   7.5165   2.0763   7.1520  -8.4977   9.5760  -7.6184
#>   -4.1707 -12.8814  -9.9301 -13.0573  -4.3214   3.3813   5.3815   5.9823
#>    1.9962  -3.9007   2.0739  12.3062   0.3560  11.4797   5.6069 -16.3572
#>    1.4082   6.4341  -8.7342  12.6657 -11.3854  16.2487   2.3478   7.0141
#>    1.0137   0.1164   1.4404  -0.5123  -6.2457  -1.4125  13.6258 -12.9901
#>    2.1578 -12.5410   7.1090  -5.6548  -0.1342 -23.7240   9.6552  -1.3177
#>   -0.9834   1.4280   8.1497 -11.6031   7.8514   1.7096   3.4896 -15.2482
#>    5.6839   2.8140   0.2340  16.2230  14.9606   0.1187   3.5830  -1.7651
#>   -0.7821  -2.0465   4.6394  -3.8735  -7.9953  -3.9239 -16.2751 -11.3365
#>    1.1868  -3.1084   4.5859  17.0607  11.3165  14.1057  -0.3135  -6.4117
#>    9.9583   2.9518  -0.7606   4.6082 -21.5887   6.4473  -0.8883  -6.3810
#>    4.0663  -4.4524  -9.3146  -2.9562   8.4336 -22.5760 -11.0995  -4.4961
#>   -1.2463  -1.7290   3.5346  -1.5529   4.7825 -12.8973   0.9904  -2.9668
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]