Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.7372   4.6328   5.5926  -2.2345  -3.8406  10.8009  14.1002  -6.0062
#>   -5.9491  -1.0120   0.7797  -9.8291   0.8864  -1.0241  -1.6519  -2.5003
#>    7.5201  -5.4463 -12.9388  -3.3077 -14.4784  -5.6324  -2.9602  -5.7958
#>    6.9475 -11.3058  -2.0085  -4.5837  -7.7840 -10.2950   2.5494  -0.8530
#>   -5.6590   3.4147  10.6884   5.8821  -3.7488   5.3377  -9.2329   6.1274
#>   -3.9574   1.6905   2.5084  -4.5981 -11.5157  -3.3068  -2.5311  -2.7889
#>    6.1966  10.7615   2.1509  -6.2699  11.4883  11.7317   8.4184  -2.5155
#>   -2.1271   5.5440   0.4896   6.3112   9.1725  -3.4008   7.1953 -12.4525
#>    7.5236   0.7170   0.8500   4.5567   3.2445   7.7978  16.8537 -10.7434
#>   -5.1703   6.9063   2.4165  -9.8243   6.4759  -3.7864  13.6161  -4.5477
#>   -6.6751   5.5971   6.7471  -6.2558  -2.4216  -6.5056  -8.5901  -8.6582
#>    1.9563  -0.6209  -6.2341  -5.5523  -3.1627  -4.3602   3.8545 -20.1129
#>   -0.7371   0.1392  -0.0481  -3.6745 -18.7724  -5.4198   8.2236  -4.0235
#>    2.6262   2.2602   3.9906  11.4829   6.8943   4.8126  -4.5235 -10.6983
#>   -2.0753   1.8156  -1.3725   5.3077  13.6391  -5.7673  -2.8156  -9.2280
#>    0.9883  -2.2134 -10.7098   6.6549   0.9091   3.8909   2.0920  -3.5647
#>  -10.9787  -6.6863   0.0173  -4.6883  -5.2768  -9.2250  -2.3450 -12.5628
#>    1.3774   1.8692  -0.7515  -0.1180  -9.8076  -2.5054  -0.9789  12.5263
#>   -1.2014  -6.9829   4.1832   2.1364  -1.8187 -15.3955  -9.9591  -5.4573
#>   11.4596   7.1422   0.2509   7.1100  -8.9286   1.4542 -11.6295   7.6464
#>    1.8945   2.5902   3.0007 -10.0442  -0.9024   3.5960  -5.9193 -14.4902
#>   -2.0418   3.1040   1.9059  -4.8431 -12.2442  -1.1955  -0.4167  -3.7720
#>   -3.3004   3.2952   7.1389   5.4808   0.4642 -10.4521  -2.8191  -1.1895
#>   -4.5241   1.9770  -5.9616  -1.2798   6.0557   2.5595   7.6885   0.6127
#>   -1.7628  -0.2457  -5.1492   8.4827  -1.4148 -10.6126 -17.6634   0.2955
#>    8.3903   3.7944 -11.3749   9.6766   8.7251 -15.3513  -5.7705   3.0995
#>   -2.4300  -2.6945   5.3980  -7.6389   1.3316  -2.1246  -2.2418   2.4587
#>   -3.1694   7.2802  -7.3651  -1.3986   9.8718   3.7231   7.1734  -5.7781
#>    2.5049  -4.6608  -6.0781   4.0247 -16.3743   0.5084  -5.5154   8.4407
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]