Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.4240   1.4932  -0.2358  -2.3355  -0.0901   7.3437   1.1076   1.7178
#>   -2.5792  -1.2333   9.2446   2.4788 -11.5579  -5.1430  13.6419  -5.4728
#>   -6.9612  -9.9717   2.1513  -5.6983  -4.3552   4.3497   6.3934   8.5972
#>   -2.3173  -0.2051   1.8719  -7.6974   6.8612   5.3454   6.3796  -3.9032
#>   -5.3525  -9.4639   8.7424  -3.7759 -27.5707  -2.8356   0.0076  -0.4529
#>   -4.3274  -9.3549   3.0415   4.6754  -7.7076 -14.8276  -3.5614  -6.1429
#>    0.6738   0.8394   0.8202   3.0833  -7.9671 -18.0351   7.2411   9.3048
#>   -1.5275   5.5868   5.1658  -0.6551  -7.5718  -2.8630   1.8181   9.1404
#>    6.2479   4.7911  -2.2466  16.8460   2.7532  -9.6917  -9.8166   2.6600
#>    3.1146  10.1711   3.2968   2.3548  -2.8576  -2.9014  24.1842   5.0666
#>   -2.9810  -7.5571  -2.2003   6.1529   3.8779  -8.6786  -8.1431   0.5202
#>    2.2769   5.0272   5.1544  10.6921  -1.9308   1.2702   3.0269  -0.4653
#>    1.6649  -1.4254   5.6606  17.2141  19.6368  -2.5211  -8.5518   9.2671
#>   -1.9111  -1.4754   2.8632 -19.6846  -2.3301   7.1434  -4.7516  -2.3644
#>   -2.4425  -0.2258  14.0052 -10.2638  -4.2697   0.5767  -2.1942  -6.5568
#>    1.0296   6.4036  -3.2874  -0.2700   9.0516   1.1768  -1.7741   6.6082
#>   -3.4931   8.7161   2.6821  -2.0298   5.9553  23.1465  14.6707  -8.6426
#>    4.7740   3.4994  -8.1523 -17.1268  -4.7060 -13.0028 -17.1750   4.4775
#>    4.6787   0.8457  -9.0898   0.7686   9.1494  -0.1541  -2.5309  -9.1132
#>   -3.9079  -3.0620   2.2763   8.3702  -5.2664   2.6447 -14.4935  -2.3403
#>   -2.1087  -1.3889   3.1254 -10.3097  -6.4070   1.3479  -3.1462 -18.0014
#>    2.2716  -5.9777   5.8950  17.0406 -12.6381   5.1494  18.9658  -3.6444
#>    0.7670   0.1938   0.4965  16.9753   5.1269 -16.0621  -1.5188  -6.1332
#>    0.3185   5.5319   1.5250   7.4396  15.5719  -1.4930 -21.9201  -6.2887
#>   -7.0484  -1.1821  14.6962   6.0630   7.3350  -0.1759   0.3869  16.9598
#>   -6.4585   5.0693  11.5892 -15.7840  16.3138  17.3379  -7.5262   3.7849
#>   -3.4192   2.0352   3.7624  -4.1861 -15.8895   5.3679 -12.6076 -15.2508
#>   -0.1721  -4.1438  -9.1698   4.3981  -3.9431   9.9248  14.5436   9.1445
#>   -0.2526  -3.9538  -4.6623   3.9046   3.4847   0.2745  15.6690  13.1897
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]