Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.9188  -2.8111  -1.8099  -9.9761   3.2225 -10.0407 -13.8979  -9.2309
#>    0.0840  -1.0227  -5.8671   3.7977   1.7426  19.1634  -3.5147  11.1089
#>    3.5966  -2.3283   1.8363  -4.3763 -10.1842  -7.9270  -2.1101   1.2355
#>   -3.1876   3.1157  -3.2982  -1.2421  -9.6976 -12.4697  -6.8536 -14.9439
#>   -3.1117   9.6163  -1.1096  -3.9863  13.6173  -8.5026   7.6991   2.4206
#>    2.4493  -8.6067   0.5305  -1.4717 -17.6288   1.3402   4.3522  -6.3624
#>   -2.0623   4.6270  -3.8241 -18.9755  20.2830 -13.0821  -3.3568   1.0948
#>   -3.9620  -0.4134  -7.5077  -4.1899   9.2942   2.3758  -0.6299   3.2084
#>    3.6912   4.6357   4.5354   8.9990   1.4732  -2.8267  -2.0441   8.0434
#>   -1.1980  -6.6816  -3.1753  -8.6331  -3.6300  -2.0620 -11.9527  -5.7416
#>    1.3663  -3.9370  -2.0759   5.3153   2.9064   0.6075  -2.6435  -2.2878
#>    2.1680   3.4707 -11.4518  -0.6525  -8.6674  -9.0004   0.6743  -6.1336
#>    0.0745  -5.8949  -2.5332  -2.7879  -8.7915   3.6804 -13.7789   4.4541
#>   -1.8269  -9.5001  -5.9933  -4.1915 -19.8888  -8.3165 -17.3140  -8.3603
#>    5.2511  -5.7261   1.2121  -1.2492   9.9218   4.6365  -2.8742  22.6450
#>    4.1443  -4.7466   3.7355  -3.4888  -0.8565  -8.2986   7.8059 -13.2502
#>   -3.0798   4.6254 -11.4602  -1.2150  12.6167  -8.5822  14.9032   6.0951
#>    1.8799  -1.1676   0.9719   9.9004  -4.4000  12.2457  -6.3252   7.6430
#>   -1.6513  -4.2563  -0.7097   8.8266   7.4434  -2.6608 -16.6996  -4.8773
#>   -0.1181  -0.9772   6.1723   2.9401  12.0966  -0.5168  12.4104  -4.0515
#>   -6.3108  -3.1904  16.2917  -4.2581   9.2739   6.6327  -1.0126  16.2166
#>    0.9876  -3.5917   5.3970  -7.6782  -0.7786   2.1552  -4.5803  10.2450
#>    4.6287   5.7948   2.0866  -2.2925  -7.3322  14.9045  -1.8674   4.4803
#>   -4.5621  -4.7558   8.5733  -5.1974   0.3421  11.1852 -14.0647  -9.1304
#>    0.6647  -4.2691  -4.3987   1.8030  -9.1747  -2.1772  -6.0359   3.8091
#>   -1.0860  -2.5123   4.9193  -9.3478  -3.0752  10.1124  -0.0068 -11.0284
#>   -1.3021  -0.0589  -0.6553   4.1769  -4.4470   6.1993  -3.2786   8.7017
#>   -0.7780  -4.7491   4.9727  -5.3176   8.1246   9.2528   8.9944  -7.2158
#>   -2.1396   0.9910   6.0503   4.3329  15.4562 -17.4982   3.7174   9.3318
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]