Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -2.3787e+00  3.5409e+00 -3.1371e-01  5.6784e+00 -4.0407e+00  1.1679e+01
#>   1.3613e+00  3.9065e+00 -7.7510e+00 -9.0381e+00 -1.2389e+01 -1.0513e+01
#>   3.7274e+00  6.5284e+00 -5.6235e+00  2.9567e+00 -4.7150e+00 -4.2317e+00
#>  -4.7613e-01 -3.9872e+00 -7.6362e+00  3.4048e-01  5.5865e+00  5.0969e+00
#>   5.2396e+00  1.5318e-01  1.6841e+00 -1.3275e+01  4.2301e+00 -1.5459e+01
#>   2.6555e+00  5.7157e-01 -8.0072e+00 -8.4959e+00 -3.9739e+00  5.4842e+00
#>  -1.0214e+00  5.1700e+00  3.8564e-01 -3.7761e+00 -1.4273e+01  3.6904e+00
#>  -2.8407e-01  6.9431e-01  1.1967e+00  1.1612e+01 -7.3271e+00  1.8692e+00
#>   2.2026e+00 -2.0764e-01  1.0057e+01  5.8129e+00  1.0609e+01 -5.4797e+00
#>  -2.5207e+00  4.1950e+00 -2.3406e+00  9.2069e+00  6.8791e+00  9.5056e+00
#>  -1.2864e+00  4.2473e+00 -8.7208e+00  4.6945e+00 -3.6628e-01 -1.0977e+00
#>  -1.7474e+00  1.8970e+00 -3.7798e+00  1.2272e+01 -2.2692e+00  2.1322e+00
#>   3.8776e+00 -5.8609e+00  9.4498e+00  2.3698e+00 -2.5399e+00 -2.0072e+00
#>   1.9123e+00 -6.4267e+00  1.0649e+00 -3.7609e+00  9.2362e+00  2.4552e+01
#>  -2.3514e+00 -1.0296e+00  1.5357e+01  4.6170e+00  8.9118e-01  1.0679e+01
#>   1.0418e+00  5.8106e+00  7.9752e+00  1.3343e+01  1.4826e+01  3.7154e+00
#>   1.8470e+00  1.2326e+00 -5.4245e+00 -7.1367e+00  9.1626e+00  2.6939e+00
#>  -2.2987e+00  9.5427e+00  3.2821e-02 -6.7669e+00 -3.2824e+00 -2.2953e+00
#>   2.9289e+00  6.3221e+00 -7.2841e+00 -2.3860e+00  1.3405e+01 -4.0540e+00
#>  -8.3795e-02 -5.3304e+00  4.8568e-01 -8.3455e-01  1.2111e+01  6.3920e-01
#>   2.5703e+00  1.0756e+01  1.5166e+00  4.1255e+00 -5.9951e+00 -1.0122e+01
#>   2.4039e+00  4.4965e-01 -5.0620e-01  6.6198e+00 -3.8076e+00 -5.6397e-01
#>   6.7714e+00 -3.7030e+00  4.1642e+00 -3.3850e+00  1.9063e+00 -5.2751e+00
#>  -9.9492e-01  7.4206e-01  3.3547e+00 -8.3889e-01  2.5759e+00 -6.2969e+00
#>   1.8355e+00 -2.1319e+00  5.3295e+00 -5.8271e+00 -9.4741e+00 -5.4342e+00
#>  -2.4370e+00  4.2608e+00  1.5018e+01 -1.7409e+01 -7.1411e+00 -1.0850e+00
#>   3.6039e+00 -6.8842e+00  2.6496e+00  8.5760e-01 -8.8551e-01 -2.0876e-01
#>   6.4855e-01 -3.2829e+00 -1.2441e+01  9.9170e+00  2.4954e+00  6.4663e+00
#>  -4.4055e+00  1.0563e+00  8.9672e+00 -4.0615e+00  2.2745e+00  8.5610e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]