Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.1469   4.6644   6.4331  18.1817   3.9744  10.5101  -5.5135  10.0988
#>    5.8198   2.4157   1.8559  -3.6720  -8.2044   0.1721   3.2691  -1.5018
#>    3.5702   1.6903   2.7643  -9.2011  -2.9558  -7.9203   9.2769   2.8916
#>   -1.1136   3.3600  11.3850   6.7524 -10.0808  -5.6578  -2.2535  -5.2243
#>   -1.2057   9.1430  10.1877   2.3819  13.8092   3.4531 -11.1109   1.3013
#>    3.1083  -2.9921   7.7781  -7.3497   0.8682   7.2036  -1.9381   9.7735
#>    0.5308   6.7499   4.3280 -14.7646   1.5462   9.2049   4.9838  -6.8034
#>   -4.8045   4.3399   6.1757  -6.0224   6.2316   2.2117  -3.8535   3.8216
#>    0.5838   0.9524   3.4798   4.6503  -3.1541   8.5678  -7.0968   9.5129
#>    0.6212   1.8737  -2.3973   3.4963  -5.8600  -3.9551  -5.1504  -4.2221
#>   -5.2182   3.8194  15.7494  -6.8264   3.5348  -8.2768   7.5462  -1.1905
#>    2.3017  -8.7634  16.8131  -9.8453  -1.0593  -8.5097  10.8263   3.7561
#>   -4.2177   3.2518   8.3659  -5.2956   0.1896  18.0529   0.1544 -15.1226
#>    3.4034  -2.3635   5.9572  -4.8475  10.9969   2.7086   8.0487 -10.4755
#>   -4.8146   8.6647   1.7499  -1.3685   7.2763  -0.4373   3.1068   2.2676
#>    3.0324   6.1377   3.2216   2.1504  -6.0979  -1.7146   3.8052   4.6414
#>    1.2588  -8.1880  -7.4340  -0.9146   0.9166   5.5232  -1.5817 -11.5741
#>   -2.6927  -4.0849   1.6922   1.6119  -8.7206 -14.3920  13.3012  -1.7760
#>    6.1114  -1.5661  -2.1272   0.8144  -1.9304  -1.3718   9.4089 -10.6799
#>    0.1726  -5.5506  11.8037  -8.5453  15.1728   8.3812 -17.7941  -5.5872
#>   -4.9816   2.6179  -0.8487 -13.6397  14.7534   7.2014   0.6561   7.1356
#>    0.2619  -0.3470   4.6967  19.0405  -3.7249   5.4584  -1.8030  -8.0542
#>   -1.3118  -9.3521   2.1261  -7.5558  -2.3003 -12.1685  -6.5816   0.5784
#>   -1.1051   1.0124  -3.2643   3.1418   9.6898  -8.3252  -5.0663  -3.0182
#>  -10.1696  -1.4406 -21.3989  -5.3194  -5.3537 -11.8768  -5.1217   0.8359
#>   -2.9200  -0.6436  12.8349  -2.0264  14.0762 -14.9025  14.6036  -1.9906
#>   -2.0262   0.2296  -1.4526   4.4852  11.7898  -0.2041   3.6338   9.6805
#>   -5.9662   5.9193   7.6321  -5.1832   7.5470  -1.0668  -1.5043   6.9529
#>   -3.2196  -8.9090  -3.5828  -0.5905  14.9915   1.4217  -7.7493  -3.4816
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]