Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.9824   0.5718   7.5896   4.3426   4.4614   3.7394 -16.0388  -1.1854
#>    1.7733  -7.6045 -17.7086 -13.7071  -4.6358 -16.4608  -7.2292  14.3242
#>    0.3032   2.7343  -4.6597  -7.2776  -4.3098   3.2016  -9.9289  -0.7264
#>    1.9979   9.2948  16.2129  12.3851   4.2771   5.1768   0.2006  -0.2567
#>   -1.6618  -0.0049  -8.6677 -10.9376  -8.4895  -4.2475  -1.0310   2.8765
#>   -3.4415  -1.5440  -6.3934  -3.8359   0.3795   8.0403  12.9098   5.5547
#>   -1.8624   4.9521   6.1610 -10.9453  15.1340  -5.3718   0.3891   1.9683
#>    0.8304  -3.3770  -4.5285 -15.0650 -16.5553   4.6365 -13.7292  -5.3683
#>    3.2935   3.4232  -8.9900  -2.5248   7.5169 -10.5980   3.1888 -11.8568
#>    2.0206  -2.0016  -2.9881   0.9160  -6.7623  -0.1729  -6.2122  15.7611
#>    4.4652  -4.9057  -9.3375   0.4637   0.2563  -3.2471  -2.5967   6.2890
#>    0.3608  -0.7252  10.5717   9.1899  11.4360   7.2036  -1.9387  10.0838
#>   -5.1576  -0.7979  -7.9352  -4.1148 -24.0594   0.2829  -5.2327 -12.6310
#>   -4.2564   0.2803  -2.2991  10.0756 -10.4723 -11.5140   5.5506   9.0681
#>   -0.5679  -3.4412  -4.0367   1.0429  -8.1792 -12.5547  -0.3229   2.6782
#>   -4.1025  -5.1003  -7.4561  -5.4914  -2.5160   0.5465   3.3305  -0.5302
#>    0.2836  -6.8997  -3.5569  -3.4634 -11.9948  -4.9269  -5.8293   6.7175
#>    4.4922  -0.3994  -0.4113   5.5931   7.5632  -8.9230  -0.3464   2.8155
#>    3.2959  10.2132   1.5692  10.6710   4.4326   6.7251  -8.3514   9.1705
#>    6.2433   4.0841   4.4363   7.9681  10.0092  -3.8291 -13.0944   7.3778
#>    3.4267  -2.6033 -10.2797  -4.4540  -3.1302  -4.0472   1.2495   4.1533
#>   -2.9671  -6.8556   2.8250   0.2966   2.4883  -6.1153  11.8498  -1.3220
#>   -5.1817   2.8812  -1.8413  -5.3970  -3.8251   0.9946   2.4399   3.7142
#>    1.8019   3.2814   8.8329   6.3868  15.5757   1.8580  13.4543 -11.6220
#>    0.6878   6.5435   9.5559   5.0030  10.1471  -1.9565  -6.5758   7.2455
#>   -4.6378  -1.9721  15.5358  -9.5409  -6.8014  11.8725 -12.1900  -1.5367
#>   -1.2479  -2.6240  -3.4915 -18.3376   5.2355  -0.7262 -13.5263  -1.7100
#>    3.7721  -6.8063  -6.9949  -0.1249   3.3159  -8.2425 -15.7249  -0.1889
#>    0.6159   1.3412  -3.4093   1.9189  15.7523 -23.2594  15.2334   0.3539
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]