Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.7413  -3.3078   2.1424  -7.9860  10.6718   4.4164  -6.6747   3.6228
#>   -0.9525   9.5668   4.7506 -10.2161  11.1849  -4.5126   7.3982  -0.8367
#>   -0.3002   4.9301   3.2637  -7.6464  -3.2062  -4.6179   3.3818  -4.6110
#>    2.4404   5.9005  -8.6653   9.8748  -3.9644  -0.1594  -2.5724   9.7599
#>   -6.9509 -10.1296   8.8871   8.6186  -4.7897  12.5594  -1.2435  -8.7750
#>   -3.7823  11.8904  -8.1491  -6.8470  22.4743   0.9746   1.2679   5.7464
#>   -2.5537  -4.0616   6.9844   4.0892  13.5282   3.7386   7.5511  -6.5854
#>    7.1050   0.1206  -3.7861  -3.5518  -4.0533  -0.2271   2.4470 -11.4416
#>   -3.4698  -7.3539  -5.1754   8.4045  -4.9722  -0.8966   3.9313   1.8256
#>    1.3468  -0.3193   0.6840  -7.1904  -7.9232  14.1407   5.5153  -9.4717
#>    9.3146   1.0147  -1.8136  -3.0943   8.2283  -3.6222   2.7446  13.5246
#>    2.9728  10.8480   4.8944  -3.2680  -5.3164  -2.4704 -12.8424  -4.4665
#>    4.2671   0.3790  -0.3930  10.7041 -10.6335  -5.2403  10.0731   0.1140
#>   -1.7838   5.4648  -9.0921 -10.3248   5.6787  -5.6305   0.9344   8.5261
#>   -3.6460   8.2323 -11.0446  -6.8950   1.7105  -7.8683   6.5592  -2.6683
#>   -0.8407  -7.1327  -6.5413   6.4758   1.6255   8.4089  11.6091 -12.0288
#>   -3.8024  -1.3269   8.7735   4.5898 -11.6868 -14.6977   6.7124  -7.5069
#>    1.3407  -7.8446  -3.1093  -2.8391 -10.8623  -3.7568  10.9897  -9.4343
#>   -0.2424 -17.1280   5.0189  -5.6006   0.1236  14.0494 -15.3945  -8.4568
#>   -4.2923   4.0188  -1.1538   3.6295  -5.0057  -6.7071   3.7228   5.6757
#>    4.4761   9.0496   1.5466  -5.0268   0.5354   4.7057   7.4019   6.0250
#>   -1.8428  10.0032  -4.3269  -3.2381  -0.2233  -9.7542  -3.2743   0.5983
#>    7.1723   3.6837  -7.5904   1.2272  -6.5757   0.5490  -1.8389  -5.9474
#>   -9.7666  -0.4508   4.7950   6.9010  -0.6245  11.8317  -2.3354   9.3846
#>    6.2706   0.8504  -4.6824  12.2050  -2.0564  -3.9020  -6.9137   7.2013
#>   -8.7833   1.9111   5.4332  -0.0884 -15.3720   0.3315  -8.1194   7.0637
#>   -3.0052   2.4827   7.6016   3.4485 -10.0805   0.4571  -5.0255  12.1685
#>    3.9457   2.2795  -8.2677  -9.7269   0.1281   5.2546  -2.0813  -1.6659
#>    4.5197  -7.0583 -14.7264   1.6873  -3.2215   3.0129   6.0260  -2.9374
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]