Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.5824  -6.7522   2.2247  -8.8911   6.8849   4.7655  11.4917  -7.9203
#>    0.4400  -2.7350 -13.5738 -21.5949  -3.2921 -17.9824  -0.1609  -1.2123
#>   -0.3422  -1.0424   7.3247   8.1146  -0.7877  -8.8004   1.5913  -5.6734
#>    6.3161   4.3443   6.9378 -13.0075   1.9896  -6.7546  -9.5280  -2.7738
#>   -1.0502   7.9112   1.2681  -3.7091  -9.0271  -2.0638  -6.1101 -11.4681
#>    0.0497   5.3240  -3.2366   8.8704   4.1991  -0.2058  11.1541   8.9305
#>   -0.6249  -9.7471   4.3020  -9.9382  -0.2109   1.0202  -5.7171   5.0036
#>   -3.0525  -5.8143 -12.0963  11.7252  -1.0003  -7.2983 -15.0752 -11.3898
#>    5.8596   0.6379   1.7704  13.9960  -3.9179   7.4643  16.6630   3.4938
#>   -0.8391   4.4325   6.9462   9.1818 -11.3296  -8.2782   8.7219 -11.5513
#>    6.2077  -7.5900   3.5818   5.8806   1.3654  -5.6130  -8.5807  -9.4568
#>    1.2679  -0.5207  -9.6594  -1.5945  21.4783  -4.5905   7.9060  -4.9781
#>   -1.6661  -5.7338  -8.4026   0.5005 -11.8238  -4.8479  17.2083  -3.7265
#>   -6.0176   4.3247   4.0690  11.5535  -9.0591  -1.7690   3.7851  14.5619
#>    3.3597  11.0268 -11.6912   0.5161   6.0901   3.0201   1.9597   0.1798
#>   -0.4627  -1.5765   6.3427  -9.4376  -5.6301  12.3668 -19.6177 -15.2252
#>    2.4565   0.4994  -5.4694   9.0276  10.4685  16.3607   0.4323 -16.4156
#>    1.2085   2.5568   2.1629   0.4938  -0.4929  -4.5691  -4.8084  -1.8439
#>    0.9269   1.3582   3.3583  18.7129  -1.4290   0.6714  11.9031  -3.5915
#>   -0.2384  -1.6044   4.6377  -3.4347   4.1837  12.9596  19.0025   3.8326
#>   -6.5528  -2.9545  -9.7639 -10.7271   6.1667  -2.4223  -6.1737 -10.9478
#>   -4.3985 -13.6431  -5.2949  -9.5287   2.3130 -14.6175  -6.2054 -11.4381
#>   -5.8149  -1.4001   2.7239   9.0197  17.5012   0.1143   6.2090  10.4016
#>   -5.6723  -2.8559  -6.6162  -5.7591 -18.1214 -11.9993  -4.1526  -0.7907
#>    1.7747   8.2972   6.4565  21.4833  20.0410  -1.8213  11.7232   5.4073
#>   -1.4488  -5.6954  -1.1172  -0.6147  -0.4623  -7.6748  -5.6041  -3.9976
#>    2.9372   1.2704  -0.9997   6.5981  -6.6037   1.6997 -10.1486  -6.3890
#>   -1.4013   3.3484  12.3591   6.1238   2.7775   4.2291  16.0938  16.7590
#>   -5.1058  10.6944 -10.1075   4.0420  20.2581  17.2744  -1.4141  22.9683
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]