Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.0374  -1.5336  -6.1149 -11.4585  -2.0237   8.5441 -11.1649  -3.3783
#>    3.7530   3.3867  -5.5748  -8.1221   5.9135 -13.1403   9.7653  -8.6281
#>   -2.8167  -1.6336  -0.6908   6.0175  -1.9288  -1.4623  -9.0049   0.4456
#>   -0.4204  -3.0950   7.1891   6.4276   1.3734   6.3890   5.5553  15.7350
#>    2.8285  -0.8039  -4.3852   2.3438  -6.6521   9.8163   1.7241 -10.5766
#>    2.0516  -5.3149  -3.9977   5.3646   5.9684  -3.1915  -7.3886 -19.4791
#>   -1.4782   6.7327  -0.1473   7.1537  -9.8860  -3.4134   2.6129  17.5811
#>   -3.0137   8.3372  -1.5870  -7.3430 -10.8916   9.7148 -11.6830  15.5781
#>    2.5023  -2.0659   9.5558 -15.2942  -2.0757  -1.5965  11.5956  -8.1630
#>    2.4908   7.8617  -2.0182  -3.9683 -16.3847   6.2289  13.6193  -0.9558
#>   -1.2967  -1.1919  -2.2035  -3.6780  -1.2615   0.9065   0.0057   7.7124
#>   -2.4402  -1.1208  -0.8555   3.4867   0.7423 -15.4172  -5.9042  -3.9278
#>   -0.5049   0.9334  -4.1580  -0.3723   1.4552   2.8628  12.0001  13.7796
#>   -3.4464   1.5088  -2.3913   1.0906   2.9585  16.4757  -4.3656  17.4516
#>   -0.9107  -9.8045  -0.9019  11.1823  13.9230  -0.9788   5.1245 -19.8611
#>   -1.4830   4.1346   1.2150   6.9783 -13.8941   0.3006  -6.4838  -2.1301
#>   -2.8210  -0.0651 -11.1119   1.4177  10.1741  -0.9395  -9.7643   1.1306
#>   -3.9553   7.1314 -11.2470  -1.6448   8.9187  15.4400   2.8976   1.2814
#>    2.9281  -1.0552   3.3932  -0.6593   9.1230   2.5120  -1.6596 -12.1605
#>    0.0195  -0.4696  -4.8472  -2.0963  -1.7001   5.5884  -6.9514  10.3803
#>    2.8084   7.2036  -0.4943  -9.0262  -4.7080   1.5875   4.9513   5.4605
#>   -3.9109  -3.2368   1.1712  -3.8027  -6.6182   4.9010   2.7824  12.3225
#>    5.4893  -2.3758  -9.9921   7.1538  -7.0969   2.3929  -4.2915  -5.3128
#>   -3.9216   4.8983  10.5917   7.9721  -2.8401   4.5244   9.0182   3.8626
#>    4.4942   1.7590  -1.5242 -11.5973  -7.8222   2.1957  -0.0923  11.7395
#>    2.0549   0.3512   5.2803   6.8448  -8.0388  -5.2928 -14.3218 -24.9549
#>   -1.8029  -1.7513  11.4075  -0.6937  -9.5854   1.1097  -4.5787   4.8125
#>   -0.4971  -2.2764  -1.0736   2.4905  13.5726  16.9746   1.8079   0.8870
#>   -0.0603  -4.0160   6.3720   1.0569  -7.1732  -3.7881  13.7301  -3.1583
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]