Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.9100  -1.2397   1.6928  -2.8398 -20.2243 -12.3919  -9.1261  14.2605
#>    2.4485  -3.1064  -0.2185   2.2441  12.5960   5.7489  -0.9597   0.3836
#>    7.0313   2.6783  -8.8775 -10.1562  -5.9640  -2.0290  -1.9640   2.4888
#>   -7.4617  -6.4382  -3.0106   5.3915 -12.0767   8.4856 -10.0667  11.3983
#>   -0.2853   4.0768 -10.9508  -1.4111 -13.6569 -11.5649   6.2195  -7.9060
#>   -3.8244   4.4693   0.3649 -14.5140  -3.0321 -13.4514   2.6171  -7.6915
#>    1.9442  10.1234  10.2097   3.2249  13.0451  -9.7466   2.2419  10.8057
#>   -1.6278   1.7883  -6.6110  13.2610  -6.0451 -16.2494  -7.2979  -7.0626
#>   -3.9086 -15.1673  -2.6387  -1.7229   2.6233  21.5778  -1.9875  17.3831
#>   -3.1630  -3.9661  -6.4075 -13.6818 -19.5680  -4.3087  -4.0381  -5.0660
#>    1.0300   4.3788   1.2204  -0.3582   9.7416  -3.9082   6.0140  -6.7553
#>    0.6318  -3.5945  16.1537   8.2657   4.7337  10.0211   6.0579   6.3888
#>   -7.2946  -5.7639   5.9167   2.1280  -0.9326   5.0593   1.4013   8.6537
#>   -7.5253  -5.9039  -4.3015  12.4216 -11.3634   1.9741   9.4937  -3.5820
#>   -0.6627  -8.1782   0.5973   4.0802 -10.3861  -1.3031  12.2828   2.5776
#>    4.2903   3.4879   2.9102  -6.7075   3.3235  -3.2867   0.2150   2.0888
#>    4.4900   6.8806  -0.4599  14.3192   3.9887  -1.3802   5.6522  -1.0127
#>   -0.0479  -5.7683  -0.1498   5.4679   6.7011  12.0204   8.9097   1.2621
#>   10.5734  -4.8804   3.6436   7.2991  -5.0086  -4.2907 -13.4022 -13.8751
#>    6.1507 -12.3002  -9.6859   4.5415  15.3843   7.4274  -4.6608  -9.9998
#>    2.6515   5.2051  -0.9041  -6.1921   3.8657   0.0111   4.8845 -18.2987
#>   -2.1500 -10.1116   1.2496   4.8682   0.7677   3.0634  -7.4915   3.2840
#>   -5.5557 -11.5368   3.5243   0.0203   2.9921   2.9721   2.5219  -2.6765
#>   -5.4225   5.1382 -11.0923   8.8642  13.9510   7.0020  -4.7262  -1.5162
#>   -2.9945  -5.1752   0.4565   5.9771  10.6228   4.2768  -0.8930  -3.2571
#>    0.5597  -5.5700  -2.0046  10.7875  -1.6586  -2.2428   4.9192  -9.1235
#>    1.5389   1.6583 -16.8568  -5.7034   2.1150  -7.3922  -9.0303  -3.6066
#>    5.6721   3.3232  -6.8244  -3.6493  -0.1284  11.9551  -2.2185  -5.9161
#>    1.8063   0.3368   4.7366   1.1419 -10.5498   6.6267  -9.7546  -3.1670
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]