Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.6948   2.7385   2.3258  -3.6766   7.2477  -0.7348 -23.2838  -2.9812
#>   -0.1922   0.5108   4.1010   2.7429 -15.2873  -0.8970  20.5136  -6.0702
#>   -1.0220   4.7308 -11.3259 -20.3873  -5.4721   6.8916   6.0073  -6.1853
#>   -9.0685 -10.9033   2.9587   7.0592   4.7419   3.7122  -3.8589   4.4267
#>   -0.5973   9.5395  -2.4456  -2.8107  -1.4171  -5.0331  15.7671  -2.8145
#>    2.2446   6.2544   1.2388  -6.1435  -7.3962  -3.1757  -7.9702   5.3838
#>   -1.6095  14.5807   2.7063  -3.8340   9.9190   7.7633  -4.0468  -1.5930
#>    0.9497  -3.7900   5.8819  -3.2986   2.6161  14.0186  17.0775   0.7018
#>   -0.0959   8.3934   5.9359  12.7524  -8.7758 -21.3502 -22.8884  -3.6537
#>   -4.4228 -10.1062   0.6051  13.9119   0.1881  -7.9986   9.5608   6.7822
#>   -6.9565   8.2409  -3.6745  -8.4878   6.7425   3.4891   4.1476  -3.8372
#>    2.8461  -2.4935   2.4231   4.4628  15.7649 -10.6794  -6.2941  16.6210
#>   -2.6679   0.0715   0.9166   4.5250 -11.9041   8.3844   0.1077   2.5725
#>   -6.7886  -3.5109  10.5798  -2.3712   4.3028   4.1786   6.4331   1.0224
#>    6.5214  -7.3308 -13.6415  -5.8342  -1.3499  -1.7033   7.8003 -14.0092
#>    1.0269   4.4341   0.5071 -11.4583 -15.4567   1.4672  18.2469  -3.5977
#>   -3.0468   8.5162  -3.4289  -2.5173  -9.3938   5.4007   3.4818  -6.9316
#>   -2.1845  -9.2216   0.0609  10.7644  -1.8674  -1.9633   2.5973   3.7777
#>   -2.7006  -0.8484   0.1801   9.3048   9.8779 -21.4559 -21.7151  -7.8186
#>    1.9417   5.4538  -3.9376 -16.4213   0.2294 -11.1355  15.8075  -5.9056
#>    7.7340   1.2012   1.2898  19.7407  22.0422  -4.3794  -9.4311  -2.9157
#>  -10.0293  -7.7860  14.4882   7.4973   0.2946  -7.9835  -5.0868  -3.2237
#>    2.7397  -3.2045   0.3665 -14.6991 -10.0212   1.9585  10.8957  11.7911
#>   -0.9597   8.6495  11.8415  -5.6273 -11.6417 -11.6483  -5.0873  -5.9569
#>    1.5481   0.0907  -7.2219   2.7542  -4.8850  17.4823  -9.1947  -2.3658
#>    4.2159  -6.6230   2.9410  -6.2258   7.2384  -4.2668  -4.6245  -1.3060
#>   -2.1275   2.8670  15.2119  -9.2699  12.2098  -9.5173  -1.6869  -0.4974
#>   -5.7960   4.7765  11.9790 -14.4819  -2.0071 -20.5256   5.4920  14.8337
#>    0.2630   5.3403  -4.6014  10.0865  -4.8894 -17.6953  17.8085  22.2598
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]