Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.8153  -8.4270 -13.1157   1.8807   1.7722   7.3508   3.3501  -4.9721
#>   -1.0124   0.0754   1.8582  -4.7820  -0.4589  -3.0153   2.2406   0.0450
#>    6.1214   0.7346   1.0334   8.0111  10.9613  -8.8082  -7.1910  -2.1971
#>   -0.5409  -1.0475  -0.4101  -3.6610 -10.2134  -2.1837  -1.6157   7.8923
#>    2.0336  -0.9012   2.6042  -5.5821  -5.2386  -2.6218 -10.3936  11.3922
#>    3.2788   6.4238  -3.2982  -7.2165   8.9121   2.7269  13.8468 -11.4536
#>    2.8911  10.0741  -2.1774   2.5091  19.6534 -14.2667  -0.6618   1.1322
#>    0.8786  -4.7877  -4.9741   2.3651  13.8431   1.2277 -13.9513  11.1296
#>    1.7741  -6.5104  -6.7832   2.4748 -13.5882   5.4863   5.0547  -9.8069
#>   -2.4306  -5.9151  -1.2627   8.8209   4.6136   2.5440  -0.7095  -9.6936
#>    2.0452  -9.3212   6.4730   6.4767  10.1169   7.0593   0.6776   0.3741
#>   -3.3698  -4.6893 -11.6878  -7.8548  12.2657  -1.2955   2.0704   4.8284
#>    3.0963   2.6404  -2.9251   2.5210   4.4728 -24.4636   5.3676  -4.6858
#>    3.0216  -0.6964  -2.0731  -2.7416   0.0394 -11.3743   0.1840  -4.0486
#>   -2.2521   6.9344  -7.9994 -14.2043  -7.2978   1.4727  13.6673  -6.7878
#>    1.4162  -7.3792 -14.7258  -3.5126   8.3703   8.0293  15.3702  -5.1456
#>    4.6347  -5.5307  -5.2736  16.9297   8.6564  -5.9274  -7.0488   3.4411
#>   -2.7455  -4.6232   9.4523   7.4941  -4.0439   5.4938   3.7629  -6.3902
#>    1.0629   2.3207   4.6792  10.5799  -3.7169  -2.4124  -8.0903  -1.4234
#>    2.7903  -2.2324  -5.8636  -7.2987  -2.7252  -1.6288  11.4659  11.0089
#>    4.7584   8.3717  -4.8836  -3.0614  10.2532   3.6604   6.2172 -10.8463
#>   -0.9072  -0.3047   0.8446  -5.9530  -6.7604   0.4086  10.3740  15.8207
#>    1.5464  -2.1715  -1.7831  -1.3004  -7.0081   9.7215   6.3541   6.0609
#>    0.3599  -3.8185  -6.9892   5.5539  12.0038  -2.3155  -5.7978  -8.4511
#>   10.9861   5.1502  -4.9071   0.6202  -3.0421 -12.5131 -12.0489  11.4802
#>   -3.0825  -5.6844   0.1064   3.4789   9.0186  18.7790  -8.5875 -10.2832
#>   -0.3051   2.3854  -2.9923  -6.0623  -4.1888  -6.9898   6.1900  -1.2808
#>   -0.3013  -0.2137  -0.2428  11.1460  -5.9054 -16.9795   3.3381 -10.4829
#>    3.6004  -0.6195  -3.2164 -19.0337  14.7028   1.8927  12.7174   2.1735
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]