Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.8227   0.8747   7.4391   4.7425  -6.3718   6.9082  25.7659  16.2236
#>   -6.8651  -1.6035   2.8807  -4.6179  13.7281  12.7265   6.2543  10.8277
#>   -0.7182 -10.3887  -8.0546  -7.0320  -1.3608   3.6029  14.3536   5.6678
#>    0.7535  -3.8424  -1.9374   3.6216   1.8049 -12.8234  -7.4284 -15.7725
#>   -9.7815   6.0903 -11.9569   3.9320 -16.2376   0.8166   9.5615   6.6312
#>    5.0553  -3.4238   1.7552  -6.8885  -0.8963   1.6307  -2.5594 -25.4071
#>    1.4415  -3.0876   9.5243   3.0388  15.2296 -12.1327 -15.8703  -5.9920
#>    0.4270   7.3376  11.1527   7.2767 -11.0910  -8.0876   4.3926   2.8611
#>   -5.8392 -11.1052  -3.5244   2.6688 -13.4104  -2.4590   6.4710   9.7437
#>   -3.3434  -2.9578   1.6931  -8.8322  -1.0370 -10.1558 -11.4084  -0.8054
#>    5.3530 -10.2598   0.0186   0.8182 -13.2735   5.9376   3.1780  -1.8270
#>   -0.2930  -9.0776   4.3291  -8.4240  -3.3933   2.4599   7.5151  -4.5738
#>    2.5381   2.6587 -11.3158  -1.5597   5.8175   4.6265  -1.7567  -6.5786
#>    1.2784 -13.8155 -11.2795  -5.0421 -11.1602  -4.8769   0.1124  27.4928
#>   -1.7416   3.0997  10.1275  -6.2011  21.5928  11.9067  -9.3973   6.9888
#>   10.7595  -1.3045   7.8618  -4.8916 -20.1712   2.8176   8.2712   7.1750
#>   -2.8077  -3.1094  23.3185   0.1756   3.4415  -3.3197 -14.2037   9.1880
#>   -4.0989  -2.2148   6.9225   0.5929  -0.6331  -8.9188  -2.9163   5.0859
#>   -7.7097   2.3741  -0.0937   0.4776 -36.8694   6.6636   5.3573   4.7615
#>   -3.4047   1.3499 -12.5194  -5.4394   1.7110   8.7620  -4.3787 -17.7868
#>    5.0298  -4.9439   0.7156   7.6969   6.7153  11.6457   9.5147  -6.1878
#>   -4.7656   7.8812  -9.0269  -1.6025  12.7284  -0.2951  -2.5671  11.3606
#>    2.0661  -6.2094  -6.3650   0.0707 -11.9886  13.5815   4.0405  22.4036
#>    3.4168   6.7986  -0.8808  -5.8088  12.5303 -11.0542  -2.4456 -13.1141
#>    7.8702  -3.8706  -0.1892   6.6065  -6.3768  -1.2112  -5.0235  -0.5058
#>   -2.3987  -3.0385   9.3624  10.2119  19.4648   6.4785  -9.6630 -12.0879
#>    3.9122   7.0486   4.2524   4.2883  -5.2131   4.0403  -1.8864 -16.0450
#>    1.4668   7.8361  -1.3183   8.8115  -3.3094 -16.6784 -15.5516 -31.1576
#>   -0.4111   3.3828  -5.0548  18.3222   0.7301   1.5179  -6.8419  11.9577
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]