Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.4235   0.2878  -2.4217   6.6874  -0.0262 -11.2560   6.6082   5.3328
#>    1.5322  -5.2968   0.8701   5.8561  14.1007   8.0900   5.2707 -22.4278
#>   -1.8593   3.7218   2.7674  -7.0241  -4.7078  -2.1096  -3.1049  -4.6317
#>    4.4967  -2.3110  -2.5185  -1.7989  -3.0036  -9.2031 -13.2654   9.9922
#>    3.7137  -2.0956   3.3277  -0.7062  -8.0936 -14.3379   7.8754   4.2505
#>   -0.2154   6.1253  -2.8669  -6.3195 -18.4725  15.6482  10.6887  10.4502
#>   -0.0813   0.3003  -2.1083   4.6747 -15.7761   4.4094   3.1035  -2.7240
#>    2.4870   3.1465   4.9427  -7.8203  -4.6675  13.3263  -9.1438   0.5382
#>   -1.1781   5.1391  -9.1152   3.0235  10.9485  16.9236  -1.8338  -5.5502
#>   -1.8483  -6.8414  -1.1116  12.1661   1.3947  20.2235  12.5616  -8.4474
#>    1.6476   3.5853  -4.3672 -16.8707   3.3857  17.8617   3.2877   6.8669
#>   -3.8686  -3.0859  -2.7270   7.3435   4.3219 -24.7858   3.9908   4.9171
#>   -0.2649   2.9872  -1.9382 -12.3552   7.4313   2.2252  -0.1183  18.1159
#>    2.9372  -5.8629   7.8042 -13.5818  -1.0732  14.4310   2.2806   4.1355
#>    2.9758  -0.0510   6.1763   4.8721  -6.0461   0.1391   1.3377  -1.7734
#>   -2.7791  -8.4573  -0.2917   0.6031  -4.0442  -1.0228  -9.1137   1.1421
#>   -1.9138  -8.3055   2.3416   3.7871  15.9567  -2.9487 -10.4832   1.4883
#>   -0.5965  -1.4523  -3.3029  -2.5223  -1.3071 -10.6619  -4.7698   7.1564
#>    3.0367   6.1815   2.7012 -15.1053 -28.3539 -10.6657  -2.5629   0.8226
#>    1.3686  -8.5010   4.5008  -5.5599  -5.8252   0.2488 -12.0070   8.9594
#>   -1.4646  -6.8422  -3.5576  12.4940   2.8127  15.5511  -0.1363   3.0854
#>    4.4546   2.0409   1.4883  -3.7529 -24.3370 -12.3287  10.1317  21.9981
#>   -0.0743   4.0471   6.7734  -6.3322  14.8364  -4.7041  -4.3602  11.0330
#>    2.2705  -2.2949  -3.8592  -4.2638  -5.1657   5.5283   0.1395  -6.9891
#>   -0.9663  -6.3047 -11.6389  -7.6179   5.1086  -6.3014   5.9518 -12.8682
#>   -3.4925  -6.0083  -0.4934  15.1517  -0.3745   3.3572  -3.6897  16.6243
#>    1.1780  -3.5153  -8.5685  -5.1810   3.4733  14.5665   9.7639 -10.0728
#>   -0.6049  -3.6207  -2.8713  -6.8173  -5.4894  -4.1157   0.8389   3.3346
#>    3.0090   0.0980   7.6039 -10.9990 -21.5325  -0.3081   8.1277   3.2715
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]