Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.4866   1.8216   3.8583  -2.5400  -7.6630   2.2250 -12.8290   1.5689
#>   -2.0346  -3.8357  -8.8207  12.2001   2.0671  -0.8828  -1.3165   1.1159
#>    2.7393   1.3232  -5.5494 -11.5671   1.3590 -20.3856  -0.5200   7.2585
#>    8.9825   0.6515  -2.2560  -1.8788   0.1371   7.9520   0.4754   3.4052
#>    7.3618   3.2617  12.3232   7.1701   5.5326 -14.3848   1.2040   1.2975
#>   -1.3442  -4.1517  -1.1213   9.0743  -7.1425  -7.2576   7.2599 -20.8879
#>    6.4966  -4.2419  -0.5973 -12.8541  12.7173   4.3852   0.0973 -11.0178
#>    0.8475  -5.4017  -9.0740  -7.1264  14.7784  12.7190  25.0345   1.1683
#>   -2.2216  -3.2227  -0.7737   1.6145   2.6893  -2.3171  -1.2580   2.9325
#>   -0.8021   1.2591   9.7955  -5.1176  -0.4441   8.6451 -11.4787  -0.4782
#>   -2.0707   5.8421   1.9531  -4.5461  12.4459  -8.5797   7.8199  -9.3489
#>   -0.0545   4.5573   5.7142  -2.9783  -2.5803   6.4004 -16.6238  -4.6055
#>   -2.5614   3.8778   9.9260   7.0848   1.8600  -2.8567   6.9139  16.0150
#>   10.2274   1.3848  -0.6835   9.4982   6.4793  -2.4636 -18.1923  -2.5294
#>    1.7219  -1.1110  10.7864 -15.4457  -8.3537  -1.6904   3.3156   6.1446
#>    1.6733  10.6959  -3.9466   7.2923  -6.2315 -13.3907   8.1190  15.4211
#>   -0.6193  -4.8326   5.6218   5.2874  -9.6151   7.9072  -2.6326  -1.8898
#>    1.5789  -0.8073   4.3614  -0.5939 -21.3487  17.6993  -6.1162 -12.4569
#>   -2.7101   6.4786  -7.4743  -1.3851  -7.4574  -3.8672   0.0589  10.5200
#>   -3.2581   4.2687  -0.9006 -14.1500   3.5721  16.5099 -10.2446  -6.1924
#>    5.6636   4.8403  -9.0097   3.2524   3.8013 -10.2591   9.7296  13.0811
#>   -2.0691  -0.7799  -1.8033  -1.1357  -1.3606   7.0147  -1.6604  -8.6549
#>   -2.8839  -4.0798   0.6127  -6.0855   8.9821 -14.1091   9.7560 -13.3854
#>   -2.1369   0.1682  -8.6993   5.3745  -7.4596 -13.0719   5.7751  -0.5517
#>    4.7243   1.1388   0.3645  14.2432   4.5201  -4.3752   1.2517 -10.0587
#>   -6.2124   2.4452 -10.1995  -5.1403  10.2539  -3.7063  -8.0729   3.8481
#>    0.2032   3.7936  -4.1187   5.3570  -7.1654   2.0132  -0.5980 -13.9143
#>   13.3146   0.7058   3.6758   1.6680  -5.8206  11.0430   9.6559  -6.7523
#>    0.5139   2.9653   5.6137  -6.1542   5.9105 -11.7130 -15.1163   5.6595
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]