Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.0346  -6.2841  -1.3165   7.0025  11.7893  -2.7810  -0.0319   6.8855
#>    6.8853  -0.4832  -2.2796   0.8673   2.9811   5.6586  -0.8290   4.0319
#>   -9.2589   5.7538  -0.9204   7.3459   7.0538   1.9784 -18.7037   6.5666
#>   -4.6213  -2.3988   5.0782   1.8171   1.4933  -5.4723   4.5774 -22.9609
#>   -0.1231  -4.4135  10.7277   3.6945   7.4943  12.0020  -2.6505   0.7373
#>    2.8777   7.5851   0.7617   4.3473  -3.6398  -0.0534  -2.8385  19.3645
#>    6.0454  -6.6087  -9.4440  -8.1493  -6.6852  -1.7977  -5.0718   1.4098
#>   -0.1537  -1.9264   5.0966  -2.3591  12.4043  -3.4766  -1.3517  -2.2068
#>   -9.0837  -0.7433  -5.2926  -7.6351   2.9082  -4.9922  -0.1428   7.1042
#>   -3.1780  -4.7976 -10.2423  -8.9963  -5.6960 -10.7425   6.5425 -10.7861
#>    1.2479   4.9875   1.9622   6.6388   5.9003   3.0893  -6.1114   9.0921
#>   -1.7697   2.1224  -0.7253  13.1049 -14.4276  13.2427   3.7475   2.3128
#>   -0.5068  -6.9679 -18.6774  -3.7379  11.0248  17.8904 -19.3567   9.5075
#>    5.0796   0.5785  -7.0678  -2.1639  -6.9436  -9.0298  10.2583  -0.3998
#>   -0.5796   8.1758   2.1101  11.4910 -12.6429   1.7435   8.9821   0.3277
#>    1.4762  12.2545   5.0594   4.1531  11.9016   7.4148  -6.7100   2.9391
#>    3.7597   4.2972  -0.2266   1.0519  -8.7590  10.1992  -5.6915   6.3546
#>    8.6132  -3.8760  20.1829  -7.3204  -3.3878   0.9247 -15.6538 -16.8579
#>    1.2914  -0.0398   7.0956 -10.6350   7.9955   5.7656   4.2322   1.3872
#>    3.3147  -6.4322   5.2836  -2.2331   0.8946  13.0180   3.3955  -6.1317
#>   -2.5393   0.0369  -3.2020  -9.2212   7.4534   0.7848  13.4753   1.1130
#>   -0.6191   2.5494   2.6890  -5.5900  19.1957   4.9638  10.4638  11.8997
#>   -3.0803   9.4295   7.8982   1.2960  -6.4891  -2.1354   4.0609   4.7502
#>    1.4558  -3.8610   0.3980   0.0479  10.6999  -3.7086   9.3034   8.4374
#>    5.7611  -5.5697   6.8518  -1.4209 -16.4104   2.6540  -4.8070  -6.0509
#>    3.6303   1.9388   3.7621  10.3741  -7.7458   8.9778  -0.3109   9.6398
#>   -0.4151  -8.6055   6.5213  -2.6612  15.1516  -1.0126 -14.2607  -9.2347
#>    2.0891   9.5349 -10.3489  -5.4833  17.5184 -21.8596   9.9112  -0.8676
#>   -0.2055  -6.7322   0.9965 -12.1426   8.2996  -8.2024   4.4764   1.5583
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]