Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -8.0188e+00  4.6745e+00  3.8830e+00 -1.3622e+01  7.9693e+00 -6.5476e+00
#>   2.4135e+00  6.5027e-01 -5.6909e+00  2.5354e+00 -4.6234e+00 -6.6281e+00
#>  -3.6130e+00  8.7384e+00 -1.2590e+00 -1.1645e+01  7.4223e+00 -8.4568e+00
#>  -3.3504e+00 -6.6645e+00 -3.1154e+00  8.3019e+00  6.0732e+00  9.5972e+00
#>  -2.6703e+00  1.5710e+00 -3.1152e+00 -4.4871e+00 -2.6620e+00 -4.8498e+00
#>   6.7976e-01 -9.4957e+00  1.9939e+00  3.9933e+00  3.6237e+00  4.5616e+00
#>  -1.0300e+00 -1.0158e+01  9.9381e+00 -4.1961e+00  1.0893e+01  6.5460e+00
#>  -1.3714e+00 -1.4972e+00  7.1722e+00 -8.8574e-01  5.5512e+00  2.1150e-01
#>  -2.3119e+00  5.7092e-01 -1.1044e+01  7.2054e+00 -4.5507e+00 -5.8014e-03
#>   1.9650e+00 -8.1176e+00 -1.1177e+00 -3.1229e+00  1.3862e+00 -9.3764e+00
#>  -1.3797e+00  5.2855e+00 -9.5797e-04 -8.1466e+00 -2.3149e+00 -2.0152e+00
#>   4.3629e+00  9.2748e-02  1.7069e+00  2.8428e+00  2.2749e+00  6.1409e+00
#>  -3.3420e-01  6.1310e+00 -3.6264e+00 -4.1134e+00 -5.0508e+00 -4.2874e-01
#>   4.9991e+00  9.7907e+00  5.4569e+00  7.5623e+00 -4.1839e+00 -8.9366e+00
#>   3.0226e-02  7.7946e+00 -4.3073e+00 -1.3244e+01 -5.1122e+00 -5.8470e+00
#>  -1.6567e-01 -1.3023e+00  6.1682e+00  4.3799e+00 -5.4653e+00  1.2660e+01
#>   3.4816e+00 -2.7248e+00 -5.2482e+00  8.9379e+00 -1.0671e+01  4.2694e+00
#>  -1.7581e+00  6.5633e+00  8.6865e+00 -3.1021e+00 -8.2941e-01  3.9957e+00
#>   5.8480e+00  9.9884e+00 -9.1158e+00  6.4815e-01 -1.4277e+01 -2.1028e+01
#>  -4.1341e+00  2.4449e+00 -1.0250e+01 -1.8493e+00  1.1023e+01 -4.1554e+00
#>  -7.0021e+00  5.2259e+00  6.0038e+00 -5.4960e+00  7.5598e+00  1.2140e+01
#>   2.0411e+00 -3.0150e+00  8.9214e+00 -2.4984e+00 -3.7530e-01  2.2090e+00
#>   4.2318e+00  4.5095e+00 -2.0893e+00 -5.6342e+00 -3.1455e-01 -1.3110e+00
#>  -1.4391e+00  5.4833e+00  3.1972e+00  9.3327e+00  1.7870e+01 -1.3209e+01
#>  -8.4854e+00 -2.1858e+00  3.7486e+00 -6.8205e-01 -2.2525e+00  1.0779e+01
#>   4.3028e+00 -9.9913e+00 -2.9561e+00  8.5468e+00  4.5842e+00 -1.5130e+00
#>   3.1784e+00  3.1486e+00  1.5434e+00 -9.3113e+00  1.8955e+00 -3.0542e+00
#>  -7.9956e+00  8.4744e+00 -5.7395e+00 -1.0478e+01 -4.8572e+00 -8.1838e+00
#>   5.8496e-01 -3.3536e+00 -4.1127e+00 -8.1614e+00 -7.4181e+00 -9.3598e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]