Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.3052   1.5834  -4.5837   6.1381 -11.0728 -17.0908 -16.7626  -5.3478
#>    3.9406   1.8907  11.1992 -10.5901  -2.5093  -9.4697   0.6438 -13.8449
#>   -4.3174 -12.4488  16.3749   9.4910  12.5489  16.8795   4.7993  -2.7187
#>   -2.4290  -1.4002   1.4027  -2.5612  -6.5243   0.7821   0.3789   3.3874
#>   -1.7356   9.4548  -2.3234  11.2071   6.0270   5.2176 -10.1530   3.1350
#>    1.6254  -4.0941   5.9057   2.9204  12.4978 -15.6447   2.9415  -7.3022
#>   -3.0487  15.2646  10.0513   1.8735  -5.6703   0.9316  -7.2733   9.4964
#>    0.4481  -6.2702   2.7981   2.1509  17.0408   8.9242  -2.2721  -1.4677
#>   -2.5352   3.7141   9.1122   4.5821  10.5415   9.2416 -20.0127   1.3334
#>    1.6121   5.6054 -17.3346 -15.3672  -3.0263 -11.4171  11.8701  -1.9186
#>   -2.5263   5.4600  13.6179   0.2430  -2.7904 -14.1900   1.4507   3.2357
#>    6.6237 -27.9243  -9.6244   5.9441   5.2605  26.9937  -5.5840   3.6687
#>   -4.7586   8.7963   2.1618   3.7114  -1.2765  -0.7999 -15.5164   9.8648
#>   -6.2119  14.9383  12.0869  -0.0285   8.0646  -9.8931   0.9332  -5.7128
#>   -0.8345 -15.5594 -21.1202 -11.6916  20.7509   7.7669   7.0344  19.8139
#>    3.2727  -4.8544   1.8746  12.0555  -2.8484   9.2706  -4.4143   2.5044
#>    0.6582   1.2200  -2.0139  -8.8894  -2.9761   6.0000  -8.3665   6.7499
#>    3.5978 -15.1892   0.9728   7.6052   8.7549  -5.9515 -16.2220   6.9245
#>    1.1013   3.4112 -17.6692 -15.1397   2.2304 -10.4852 -12.2337  17.2254
#>   -4.5356   2.5122   4.3019   9.9546  -0.6406  -5.9611 -21.6922  -2.7619
#>   -0.2231  -5.5092   5.6141   0.4538  -6.8529  -5.3510 -18.7863  -2.8748
#>    3.0550   1.3467  -6.4608  -3.0257  -2.1310   4.1361  -3.0521  -2.8458
#>   -3.7644   7.0262   4.7741  -5.3720   4.0487  14.4133  -2.8319   0.3900
#>    1.1639   3.2431  -6.5921  -9.8217   8.5429  12.5020  15.7258   4.5391
#>   -1.1710  -4.6705   5.9731   1.2742   4.1296  -3.1673  17.1912  -8.5328
#>    2.1463  -7.7196 -11.9454   3.6759 -11.6863   8.0377  -2.4014   4.2443
#>   -2.0626  -0.6780  21.2331  -3.5198   1.2828   5.7143 -14.6567  -1.5910
#>   -0.0783   3.7810 -10.9354  -7.7134  16.1776  -0.5663  10.4791  -1.2474
#>    1.3320 -10.3248   8.3930  -1.5960   0.7677 -25.3169  11.1711   3.7886
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]