Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -6.2101   9.1185  -1.8670  -6.4591   2.5618  11.3172  -4.3821   3.1105
#>   -2.4393  -6.8816   3.2293   8.5030 -11.8052 -10.0143   3.2133  -1.7253
#>   -1.8052   0.7470 -16.3309  -5.8964  -5.1758  -8.3300 -10.7493  10.9457
#>   -3.2727  14.3368   8.4960   6.0826   0.7161   5.2090   6.9864   6.4386
#>   -7.2670   7.5826  -5.6343   6.9464   4.4122  -0.7374  -5.0015  -1.4322
#>    2.1807   1.1508 -13.4206  -3.3896   7.3074   5.4382   0.7846   4.8138
#>    6.8440   3.3460  -0.7076   1.9999  -2.9588   3.9588   4.5887  -8.2706
#>   -6.4947   3.4757 -10.4605   2.9481   1.8951  -3.5458  -6.0539   8.2208
#>    1.5339  -2.1833  -2.3324   1.1552  16.3869 -10.2488  -4.3299   8.8177
#>    3.6386  -4.1725   1.6154   3.3637   2.5686 -20.3112  -7.2147  15.4925
#>   -3.9754   7.4672 -11.0007   3.9904  -2.1636   2.6294  11.0357   9.1544
#>    1.0825   5.0482  -6.6633   0.3296  -4.6025  -2.3428  -1.5126  13.2328
#>   -2.3841  -5.9685   3.5792  -1.1213   5.0920   3.7071   6.4324  -9.1056
#>   -3.4894  -3.2852   1.7646   6.7690  -3.7430  10.4152  -5.9921   7.3134
#>    1.8703 -16.2415   1.4007   5.1320   5.5489 -13.0178   6.8390  -7.6221
#>    4.0410  -0.2403  -0.2433  -0.8852   2.9912  -5.2536  -0.7030  18.3249
#>    4.1349   7.7658   7.0653   0.5951  -4.4099  -1.0293  -6.1531 -15.4617
#>    3.9520 -11.4874   5.3329  -4.3043   8.0465  -0.4940   3.1498  -5.5065
#>   -0.2038  -6.0580  -5.3845   0.2241   7.6020   9.7723  -7.3332  -6.8052
#>    1.2361  -4.6275   5.1346   6.8869 -20.4626  -3.8403   8.8436  12.9356
#>   -7.3041  -1.9487  -3.8306   3.7365  17.2030  -1.5172  -1.7095   0.1073
#>   -6.7670  -8.4325  -4.7693  -2.3777   1.8015   0.3823   2.6246  10.2152
#>   -7.7016   0.5234   4.7479 -12.5596 -16.8053   3.5840  -6.1286   2.4994
#>    6.0311   3.8515  -2.3541  14.1136   7.2072   4.1987   1.5736   3.1507
#>   -5.0376   7.9029  -0.3829   2.5462 -14.8511   8.2748  -8.0424 -12.6077
#>   -4.6770 -20.3535  -1.2235  -5.4131  13.3542   3.0645  -0.6015  13.3827
#>   -3.4830   6.6595  -1.9712   3.9635  -4.2631  17.0630   1.5010  -9.1325
#>   -0.0670   3.3956   7.6852  11.6633  -1.8617 -10.2024  -9.9527 -16.6496
#>    5.8974   0.4273  -5.0006  -9.8298  -4.8006  13.3415  -2.2487   0.4192
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]