Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.3363 -10.0419  -2.6670 -18.2367  -6.9532  -0.7196  16.3299  15.8587
#>   -2.6242  -2.4471  -2.2343  -1.1525 -13.0694   5.4759  17.7988  -4.7645
#>   -4.1170  -3.9966   0.6953   5.4297   5.1772  13.9072  -0.7007  -5.8837
#>   -9.9598  -5.9400   3.2368   4.7242  -4.3028  -2.2740   9.8286  -9.7897
#>   -5.6022  -7.9602  -7.9000  -2.8328   6.3413   3.4274   8.4188   1.5803
#>   -3.4809  -8.7980  -9.1923  -9.5750   4.9035   4.2246   4.8785  -7.1431
#>   -8.4140 -10.0009  -5.4753   5.2253  -4.0068 -13.6264  -4.4413  16.9537
#>   -0.1722   3.1110  -3.7797  10.7461  10.8417  -1.3801   0.1622  -8.7147
#>   -1.6787   7.7871   2.8406  -9.2262   0.9129  -3.1607 -11.2113 -20.0364
#>   -0.3862   0.8748  -5.1719  15.4512   4.1317  -6.0348   1.6015  -5.4444
#>   -2.1833  -6.0798  -1.1479   5.4557  11.7432   0.2242  20.9796  -1.0657
#>    3.6630  -2.2125  -9.2257  -8.6615   4.1665   0.9297   4.8879   1.4074
#>    2.3128  -1.2364   7.7583   1.8383  -0.5645 -13.0141  -3.8985   7.5363
#>   -1.0963  -6.0368   6.1886   4.4570  13.7579  -2.9187  -5.8656  12.8506
#>    7.4092  -0.3004   5.9274   3.4956  -4.6605   7.8882   3.1844   1.1576
#>    0.5059  10.4921  -2.7905  -6.8412  10.3135   4.6601   3.8002  17.0318
#>    4.0664  12.8884  -6.6000 -13.2856   8.1711   5.4287  -3.5988   8.2070
#>   -0.3659  -6.5509 -21.2951 -10.0726  10.5397   6.8938  -0.4340  -0.2986
#>   -7.1746   0.8012  11.7207  12.0253 -11.6830   8.8723   5.7269  11.1269
#>    1.1422  -7.8123  -9.7902  10.3567  -7.8402  -8.9201  -1.2143  -3.9008
#>   -0.3362   2.1308   7.1742   1.3936   1.8426  -9.8587   8.2590  14.6265
#>    0.1819  -2.5455  -4.7871  -3.4159   1.1901  -6.8771   7.2125   3.9897
#>   -4.9609  -4.6204  -0.0886   9.1302 -11.5868   7.6954   2.2312  -3.4116
#>   -5.9213  -8.7601 -15.5261  -5.1006   1.8575  -3.2237   7.9049  20.0632
#>   -0.3581   8.2605  14.6275   1.1350  -8.0672  -8.1972  -6.3011   8.7307
#>   -3.8978  -3.3300   2.7334   9.1648  -1.3268  -7.7717  11.2868 -11.5360
#>   -1.9044  -4.1962 -13.8932 -21.6736 -20.9493  12.5132  15.3223   2.3947
#>   -4.2702  -0.6774  14.4121 -10.4399  -4.7489  -0.0565 -17.4295 -18.4775
#>   -0.0247  11.1074   8.9426 -12.3020   1.4101   6.6588   8.1552 -12.6988
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]