Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  1.8792e+00 -5.3858e+00  3.3288e+00 -5.8907e+00  8.8195e+00 -4.6321e+00
#>  -3.6146e+00  3.8169e+00 -4.9913e+00  1.3466e+01  1.5333e+01 -1.4301e+01
#>  -1.0193e+01  9.7816e+00 -1.0996e+00 -8.5520e+00  8.4105e-01 -4.3648e+00
#>   1.5094e+01 -1.0583e+00  1.4660e+01  2.6713e+00  7.9599e+00  1.6390e+01
#>   7.8907e+00  2.0925e+00 -1.5065e+01  1.2973e+01 -1.1026e+00 -2.7301e+00
#>   1.3030e+00 -2.3146e+00  7.8115e-01  3.1535e+00  2.5843e+00  2.6675e+00
#>   4.9542e-01 -6.3515e+00 -3.3054e+00  6.1620e+00  2.4367e+00  3.4734e+00
#>   2.6834e+00 -4.2956e+00 -8.2819e-02  5.0317e+00  1.8823e+00 -9.9065e+00
#>   1.8778e-01  4.6141e+00 -2.2498e+00 -7.3048e-01  3.1028e+01 -6.1658e+00
#>  -2.6981e+00  1.6428e+01 -4.4325e+00 -2.0331e+01 -2.9794e+00  3.0111e+00
#>  -1.7363e+00 -5.1245e+00  9.6349e+00 -2.4382e+00 -8.3971e+00 -1.1142e+01
#>   5.7344e-01  2.8591e+00  2.9847e+00 -1.1625e+01  3.4689e+00  9.8731e+00
#>   8.6056e+00  5.2444e+00 -2.2725e+01  7.1959e+00  1.3686e+01  4.0357e+00
#>  -1.1577e+01  1.5868e+01 -7.5000e-01 -6.1847e+00  7.2168e+00 -2.6549e+00
#>  -1.2776e+00 -1.2964e+00 -1.2659e+01  9.7742e-01  9.8535e+00  6.5964e+00
#>  -3.0673e+00 -1.3654e+01  6.1103e+00  2.5727e+00  7.1996e+00 -7.4975e+00
#>   6.3611e+00  3.5280e+00 -6.5242e+00  7.7940e+00  6.2022e+00  5.8364e+00
#>   3.3390e+00 -8.0657e+00  1.1978e+00  4.8404e+00 -8.8597e+00  6.4237e+00
#>  -6.1187e-01 -4.0615e+00 -2.2535e+00  1.0399e+01  1.1769e+01 -4.2436e+00
#>  -1.4007e+00  1.2731e+01 -6.3327e+00  3.2181e+00  7.3150e+00 -4.2840e+00
#>  -5.2509e+00  2.0465e+00 -1.3589e+01  4.6183e+00 -7.2462e+00  1.7004e+00
#>  -7.6813e+00  5.8122e+00  2.3021e+00 -2.5548e+01  7.9606e+00  2.8913e+00
#>  -2.6242e-02  4.7862e+00  1.0873e+01  1.8983e+00 -3.9481e+00 -7.3546e+00
#>   5.1275e-01 -8.7306e+00 -6.8629e+00  1.5566e+01  2.6107e+00 -1.8449e+01
#>  -8.7155e+00  1.1945e+01  3.5693e+00 -2.1491e+01  1.7815e+01 -1.2563e+01
#>   2.9699e+00 -1.0409e+01  7.8728e+00  1.2377e+00 -1.0165e+01  2.2859e+00
#>  -6.1068e+00  7.7523e+00  1.3736e+00  1.7710e+01  9.3003e+00 -1.8531e+01
#>   4.7490e+00  1.4199e+01 -1.6164e+00 -1.2385e+01 -1.8251e+01  1.8897e+01
#>  -3.5350e+00  2.3773e+00 -5.8177e+00 -2.5659e+00 -1.1824e+01  1.2248e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]