Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.7250  -8.4282  10.9705  25.1556   3.9656  -3.9843  -5.4015  -3.3493
#>   10.5694   5.3795  -6.0322   8.2266 -13.3990  -2.8590  -0.7411  -8.3572
#>   -3.9332   1.8571   4.7006   8.8207 -12.8341  -6.3409  -9.3062   1.6715
#>    4.3418   5.0598   0.1717  -1.2498  23.6345   6.2150 -19.5062   6.2430
#>    5.8265   4.9091   1.1509  26.4678  -3.7608 -11.4816   1.1317   5.1793
#>    2.6473 -16.0673   3.1001  10.3443  14.1546 -13.9760  -0.2670   7.7266
#>    3.7014   9.4910   2.3459   3.1847  -9.2104  15.2986  -0.8423   6.5073
#>   -2.3076   1.1558 -18.4287  -4.0526   0.8455 -16.9122  -7.5676  -0.9610
#>    1.9849  -7.1740   9.9974 -10.0709   9.8181  12.5892  -0.6088  -7.3053
#>    3.0574  -1.8354  -6.1069   5.0326   6.3530  -2.9266   4.3755 -12.1816
#>    3.4249   1.3292   3.7645   0.7606  -3.1900  -2.2632   4.0344   5.4776
#>   -9.7556  -1.6295  11.1228   7.6204  -2.5429  -1.2315   9.7837  -0.7951
#>    4.2996   6.5658  25.4355  16.4203   8.8159 -15.3252   5.8350   4.5160
#>   12.6958 -11.5574  10.3848  -1.1438  17.8979  -8.9263  -9.1516  12.6326
#>   -3.3983   1.4272   1.7906  -3.1975   6.0251   5.1568   5.3156  -1.0467
#>   -6.9068  -2.7230   1.3269   1.8554  -0.5369  10.7842  -1.8818  -3.8361
#>    6.4304  -5.1225  -2.7330  -3.2588   7.7150 -15.2676   0.2653  -0.5105
#>   -0.8296   5.9223   2.3489   0.1923   3.2507   5.2096   8.1143   2.2531
#>    2.1810   8.0862  -8.3544  16.5000 -18.4632   4.7049   6.5420  12.7624
#>  -10.4374   5.6388  -5.2228  17.7059   6.9261  -3.9237  -1.4999  -9.2772
#>    3.2019   0.6588  -1.8807  -6.2926   0.1557  -3.3038  -5.4629   4.7243
#>    7.9428  -7.1186  -0.4914   3.5046  -7.7914   5.7057   8.7280   3.8084
#>    4.9315  -7.4707   1.4627  11.4521   0.7455   5.3401 -10.0684   5.8012
#>    0.3482   5.6267  -0.8318  13.1464  14.3252  17.6890  -0.2677   4.9739
#>   -0.3150  -5.6911  17.9565  -6.1935  17.2241  16.3586 -11.3513  10.8611
#>  -14.2593  -1.5049   4.7490   3.3280  -7.3528  -1.6771 -13.2362   0.8403
#>    0.7245   2.9692  -7.7825  14.0206  -3.4675   3.9569  -9.0859  14.2275
#>    2.8282  -4.2137 -15.2626   3.7467   3.1872   0.0565  -4.8512  -8.3182
#>    5.7686   3.1803  -0.0143 -12.7116 -26.7579  -3.8621   7.4107  -3.6132
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]