Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.1794  -3.9381   7.0735  -9.2118   0.9546  -4.6239   8.5634  -6.9773
#>    1.9154   1.8913   7.5092   0.5552  -9.9346   6.4010  -1.6639   7.7503
#>   -2.6646   3.1313  -8.2212 -13.1989   7.6207   4.0096  -8.8543  10.1340
#>   -1.6313  -5.3401  -4.1557  -6.1677   7.5252  -1.1533  -6.8553   4.4996
#>   -2.0358  -2.5553  -1.9135  -5.1754  -1.6110  -7.1631   4.9943  -5.0803
#>   -0.6328  -1.0509  -8.3323   3.7066   1.0714  -3.7327  -9.5693 -12.4113
#>    1.8463  -1.0055  -5.1535  -8.6584   9.4835   5.4466  -5.1379   4.2689
#>    0.8820   2.5867   5.6925  12.5437  -3.7519  -7.4582   1.0570  12.4213
#>   -1.4760   0.6808   4.4194   2.7002  -6.0580 -11.7653   8.4315  -3.5401
#>   -1.4007  -1.9311   4.8063   0.9439  15.6114   0.7082  -7.6117   5.5618
#>   -3.5472   3.3041  -2.6917  17.1902  -4.0781   4.7791   3.9214 -11.6281
#>   -0.7346   3.8247 -13.9388 -13.9550 -11.9220  -4.3565  -3.8134  -4.7469
#>   -1.2331  -2.1906   9.6243   3.3334   4.2680  18.3745  -3.0787  13.6531
#>    2.7895  -1.9506   7.6748   2.7703  15.1214  -3.2748 -20.6841  -5.6664
#>    1.5247  -0.3030   3.4951  -7.4808   5.9173  -4.7509  -7.3708  10.5828
#>    3.0584  -1.3683   0.0408   1.8690  -5.2394 -12.3501   5.7697  16.9476
#>    3.0622   1.6548  -1.5128 -10.3283   1.4842   1.9541   7.7540 -10.0913
#>   -1.7088   2.4168  -0.0358  10.4642  -1.2960  -7.4867 -14.7129  -2.0177
#>   -0.1353  -5.5485  -7.8592  -2.8101  -1.8221   6.3026  -3.1771  -2.5096
#>   -0.1605  -5.9435   9.7720   5.9767  -5.8232  -2.0531   5.5221 -12.6244
#>    5.4104   1.4104  -8.1802 -11.2445  10.9029  16.0497 -14.3194   0.2794
#>   -2.9107  -4.9550  -2.7400  -3.8605   1.7396  22.5797  -7.0195   1.0087
#>   -1.9935   6.3472  -0.2343  -7.3847  11.9060  10.6991  -2.4574  11.6684
#>   -1.3092   4.5347  -1.9084  -6.1592  10.0787  15.7895  -4.3406  -4.2463
#>   -4.5820  -8.2995   3.5279   4.4152  17.4065 -13.1182 -16.4217  -9.6643
#>   -1.4577   4.7303  -7.2366  -7.1705 -17.8069   4.6852   2.1436  -5.4950
#>   -3.4177   5.6711   8.5124   5.3524   1.3937  -8.1730  -7.1572   7.2018
#>   -0.7085  -8.7912   3.0505   2.2731  10.2857 -17.7151 -10.3583   7.7587
#>    1.4611  -0.8549   2.5630   7.3177   0.0503  25.4050   0.1536  -7.8269
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]