Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.8268   1.7566  25.2124   7.2261   8.2564  12.6092 -14.2583 -11.5287
#>    6.1669  -3.0821   6.2523   4.1238   7.3594   5.3705  -7.0274   6.6871
#>   -8.2146 -12.4772  -7.5117   7.1113  17.3588  10.2529   5.6448   4.7558
#>   12.6773  -3.2798  -6.2097  -2.6679   8.0700  -9.9351  -6.9065  11.9815
#>   -1.1164  -2.9251   5.9107  11.4200  12.9925  10.0514   7.2893  -2.3667
#>   -7.0482  11.1797 -11.2108  -0.8015 -18.6916   5.2554  14.1655   0.9194
#>    2.0216   1.4207  -8.5305 -16.1985  -6.6315   1.1138  -0.9759  -3.1522
#>   -5.8327  -6.1287  -2.4522   1.8951  -4.5400  13.7297  -0.9638   3.1605
#>   -1.4188  -3.6991   5.6835   2.2708   2.5827   1.8319   0.1044  -2.7261
#>   11.4982   1.7019  -0.4870  -5.3422 -14.5557  -7.3097  -3.7626  -6.5087
#>    6.9863  -4.1417   0.4567   7.7888   6.5228  -1.1557   3.1177  13.2544
#>    2.4977   5.1740 -14.0243  -8.8318 -11.2456  -0.0797  -5.7639  -8.9322
#>   -6.9579   5.5570  -7.6725  -7.7461  13.8574 -24.0545   2.4006  -7.4875
#>   -4.1015   2.0995   0.6497  -3.4877  -4.5932   5.8009  -8.7424  -5.9722
#>    1.9630  -8.4907  -3.0012 -19.4748   0.6332  -2.1485  -4.9878  -1.1321
#>   -3.5631  -3.3813  -5.4723   3.7823   5.1525 -12.1421   0.2567   0.9103
#>    5.5368   6.4795   5.8897  16.9116 -10.1002   9.8383  -8.7364 -13.9280
#>    1.5624  -8.5233   9.5308   3.5206  -6.5621  -1.1458 -18.0790   4.6674
#>    4.8489   2.2027  -0.2048  14.5968   8.5715  -3.0682   3.7143   1.3081
#>    0.8422  -3.5423   8.5055  11.9697   4.2253   3.9077   5.3860  -6.8834
#>   -5.7921  -2.8770  11.9408  11.7675   6.0040  -6.0863   0.2983  -2.7187
#>   -1.9578 -12.4719   5.7096   4.2805 -10.0009  -1.6440  -2.7296  -9.1526
#>    1.9222  -5.9536  -3.1811 -13.2929   2.2943  -1.8315  -2.2698  -5.1075
#>    8.7532  -6.1764  -9.5641 -18.3473 -12.2465  -5.0860  -3.9357   2.5006
#>  -13.9150  -3.4904  11.8376  18.1129  -8.1957  -2.1573  16.4605 -16.8901
#>    0.9289   5.5585  -0.7901   7.0534   4.8403   3.3536  12.5340  -3.3383
#>    3.6373  -7.7348 -12.6885  -6.5279  -7.6778  14.8828   6.5296  -6.8275
#>   11.1622  -1.9410  -0.2101 -11.7588   2.7983  11.0770   1.1867   2.5921
#>   -1.8506  -4.0477  -8.8516   2.6037  -4.3581  -0.6596  -9.2576   7.2431
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]