Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.8245   5.2668   6.8981  -4.5933  -2.1690  10.9942  -1.5282   7.7898
#>    3.4214 -10.0369 -12.0690 -12.0094 -12.9308 -20.6070   0.1269  -3.5434
#>    1.3614  -6.2832 -10.2663   2.0378   2.7738  -6.7335  -0.4032  -4.7811
#>    1.9838   2.7140   6.0424   7.5026   0.3010   1.1168   1.9950   4.9336
#>    5.1549  -1.6413  -1.3271   3.6030  -0.1130  -3.1924  -7.3464   9.9467
#>    5.0094  -1.1081   8.8214  -2.1542  -4.2967  -7.5526   7.0349   8.7296
#>    6.1922  -0.6049   0.0349   8.1172   6.5177  -3.8198   3.5903  10.6680
#>    5.8254  -4.0215   0.4075  -9.8716   9.9498   6.0882  25.4013  -7.2722
#>   -2.4909  -3.7183   7.2182  -1.7926  10.0069  -8.6853  10.3960   0.7976
#>   -2.3247   6.0328  -6.8645  -0.8955   6.4752   3.2846 -11.8066   3.8223
#>   -0.3913   5.5591  -2.6213  -3.4559   2.3704  13.8402  11.5054   3.8928
#>    0.8651   2.0450   5.2707  10.0426  -0.3397   2.5584 -10.4950   0.6790
#>   -3.0289  -4.0275  -0.5551  -3.3836   4.6235  10.4512   0.6820   2.6848
#>    2.3565   0.8260  -3.4258   5.6239  -0.4688  -9.6714 -10.2801  -2.2470
#>    2.8876   2.6073   6.0162  -2.3681   8.8298  -9.1912   6.1130  -4.5827
#>   -1.3057   3.4448  10.4644   4.8420  10.2680  -8.7003  -5.2126  -5.7616
#>   -2.7433   0.3315  -8.8132  -5.7281   6.5782  -4.3726  -4.0937  11.1725
#>    0.8435   1.4053  11.0125  -1.3573   1.5665  -4.5346   1.8229  14.5148
#>    1.4586 -11.3073 -14.3211  -1.1334  -2.4172  -4.3308  12.2667  -4.1112
#>    3.4125  -5.2829   9.2569  -2.7894   4.3900 -15.1102  16.2394   1.1730
#>    1.1316   2.9915   5.5521  -2.8638   8.5938  15.0421  -1.6030  -4.2427
#>   -0.1688   0.5495   2.9811   0.5726  -6.5691  -6.9999  17.2904  13.2416
#>    0.0614   0.2379  -2.9278  -4.5077   2.0333  -6.8194  -2.2769   6.8851
#>   -0.7966  -0.1735   8.7751   4.5213  -3.8701   4.1515   6.7270   7.8442
#>   -0.3308   3.4538  -3.2838 -16.7542  -7.0407 -15.3083 -16.6815   1.1446
#>   -0.9612  10.2966   2.8524   4.4385  -5.5516  10.4120  -7.3594  -8.9473
#>   -5.4615   1.2501  -7.3209  -4.1472   3.9163 -16.6378  18.3221 -17.1170
#>    1.4738   2.0108   3.6077   5.1772   9.6229  10.2336  -7.5156   2.7033
#>    0.2844   6.2825  -8.3219   3.2492   0.6667 -10.2333  -0.3739  11.2053
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]