Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -2.3438e+00  2.4850e+00  1.6594e+01 -3.4298e+00 -1.1322e+01  2.5386e+01
#>   2.1199e-01  5.6819e+00  3.1402e+00 -3.2158e+00  5.3946e+00  7.7262e+00
#>  -4.8189e+00 -4.5753e-01  1.9808e+00  1.2439e+01 -1.5843e+01  2.8460e-01
#>   5.8777e+00 -6.0700e+00  4.9891e+00 -9.0818e-01 -1.1609e+00  5.4806e+00
#>   5.0447e+00 -1.3098e+00  1.1204e+01 -4.7528e+00 -1.5030e+01 -3.5499e-01
#>  -9.0119e+00 -1.7834e+00 -3.2071e+00 -1.4537e+01  1.0271e+00  4.0520e+00
#>   1.7639e+00 -5.7360e+00 -7.4024e+00  1.0868e+01  1.9365e+01 -1.4467e+01
#>   1.2370e+00 -2.3708e+00 -7.9496e+00  4.9205e+00  8.1970e-01 -1.0414e+01
#>  -1.2005e+00  8.2899e+00 -9.0001e+00  5.6647e+00  1.2526e+01  1.4225e+01
#>  -4.8066e+00 -8.1628e+00 -5.3962e-01  4.1076e+00  7.7064e+00 -8.0497e+00
#>  -3.6532e+00 -6.7250e+00  2.7122e-01 -2.9055e+00  1.5755e+01  5.2388e-02
#>  -2.1206e+00 -3.9653e+00  7.6236e+00 -2.4668e+00 -8.7052e+00 -7.5951e+00
#>  -1.6382e+00  1.0314e+01 -4.0975e+00 -4.3876e+00  1.8999e+00 -5.8774e-01
#>   4.5558e-01 -2.9188e+00 -5.8258e+00 -1.2185e+00 -3.3196e+00  3.7037e+00
#>   4.3954e+00  2.2439e+00 -1.8564e+00  1.4197e-01  6.4288e-01  1.7669e+01
#>   4.8130e+00  1.6600e+00 -5.3562e+00 -7.9489e+00  1.2706e+01 -2.7425e+00
#>  -2.8657e+00  5.3949e+00 -2.9963e+00 -7.6891e-01  3.9105e+00  8.3957e+00
#>   9.7749e-01  4.2285e+00  4.2728e+00  2.4541e-01  8.6050e+00 -4.4170e-01
#>   4.6693e+00  3.2152e+00  5.3720e-01  1.3279e+01  9.6078e+00  1.1393e+00
#>   8.2061e-02  5.2474e+00 -9.8109e+00 -1.2072e+00 -1.0195e+01  3.2099e+00
#>   5.5098e-01 -3.7260e+00  1.0820e+00  6.9889e+00 -9.5013e+00 -1.2087e+01
#>  -6.0613e+00  1.4178e+01  1.9362e+00  4.2236e+00  2.3609e+00 -4.7164e+00
#>   1.4139e+00 -4.8287e+00 -1.4900e+00 -2.0428e+00 -1.9353e+01  1.1550e+01
#>  -2.9693e+00  9.7311e-01  2.9526e+00  3.1726e+00  1.6335e+01  4.4718e-01
#>   3.9060e+00  3.2903e-01  3.3510e+00  1.2291e+01 -5.8344e+00  1.2486e+00
#>  -5.6685e+00 -2.6804e+00 -8.3612e-01 -1.4761e+01  1.8184e+01  3.8310e+00
#>   7.2578e+00 -4.3200e+00 -1.0552e+00  6.6765e+00 -2.5970e+01 -1.2665e+00
#>  -3.3254e+00  7.6203e+00  1.4048e+01 -8.6996e-01  1.0463e+01 -4.7697e+00
#>  -3.7995e+00  9.6784e+00 -2.2545e+00  5.3848e+00  2.3302e+01  5.7294e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]