Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.3056  -7.8159  -6.4075   6.1613 -11.9068   2.5058  -4.7512  16.4364
#>   -1.9584  -0.8362   7.1764  -4.5980   4.0851   4.1650  10.9455  -5.5731
#>   -0.0496  -5.8116   2.4900  -3.8716  10.0556   2.4398 -15.1719   0.2124
#>    0.1863   1.4993  -7.7411   2.7104   6.5400   5.8112  -4.4483   3.6627
#>   -0.1585  -4.1472   0.0483   7.5303 -11.1412   9.2981  -1.9070   4.2120
#>    2.1278  -6.5877  -2.4971  -5.9609  -6.2670   4.7855 -18.1277   9.0975
#>   -7.0960  -0.2551  17.5011   9.7086   2.8606  -7.1148   0.0213  -4.6231
#>   -2.6878   4.6929 -11.5868 -10.0519  -8.2063  -5.1566  10.9535  -3.7494
#>   -3.2860  -9.0068  -8.0979   7.7353  -1.7326  -3.9381  -9.5958  -9.1043
#>   -3.8830  -3.9193   1.4615   5.4842   3.2720   4.8428  -4.9542   0.4090
#>    1.9859  -0.3492  -7.4509 -12.4787  -0.4488   5.9391   2.2614   0.3480
#>   -2.1591   1.8424   0.5409   8.8389   7.5470  -9.6862   0.5086   2.1500
#>    7.0653   5.1423  -4.5526 -11.2576   0.3406  -0.8729   2.6633 -10.2472
#>   -1.3504  -0.5707  -3.2394 -10.6178   1.2861  -7.5086  10.4821   2.4296
#>   -1.2625  -2.2756   6.4089   1.3783  -0.7172   3.5674   8.3333   3.8049
#>   -3.8014   6.5931   8.3853   4.3731   6.2211 -14.8918  -1.1326  -9.6734
#>    3.3392  -1.6871  -0.6311  -5.9305   1.1477 -13.6541   2.0424  -7.3964
#>    5.2311  -4.4934  -8.3005   8.0256  -9.3237  13.0139  -3.5230   4.5461
#>   -2.2072  -7.5747   3.9764   2.1093  31.7618  -9.3007  -9.1003  15.0032
#>    4.2302   0.9131   1.9145  -4.7453  12.6869   5.4029  -5.1272  -3.6743
#>   -4.4550 -12.3645  -9.0223   4.2341  -4.5045   8.8591  13.8550   3.3347
#>    1.8700  -3.3781  13.0072   6.4834 -15.3154  -8.7369  -3.8119  12.5658
#>    3.0498   8.6054   5.5034   8.3305 -11.9736  -8.8896  -9.3150  -0.5306
#>   -1.2353  11.5427   6.3127   5.6692  -9.8486  -5.4571  -0.0794   7.2823
#>   -2.1391   0.2026   0.3323   0.4729   8.4796  -3.8518   5.4850  -0.5194
#>   -4.6539   5.0152   3.5524   1.4340  -6.1565   2.7737   2.9640   7.1451
#>    3.9821 -11.6528  -2.9261   1.3423  -0.2858  -1.1649   9.2995   7.5925
#>   -5.7471   1.9885  -4.9137  -0.4508  -1.2386 -11.6626  -3.3683   1.9939
#>   -2.1255  -3.2957   4.2624  -2.4059  -1.6073   8.8491  13.0591  -0.6741
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]