Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -6.1694  -7.7452  16.9519   3.1878  14.0328   2.0770  14.9792  12.0712
#>   -0.0814  -1.3910  -4.7454 -15.9979 -16.7072 -14.6367 -16.3910 -10.7704
#>   -5.2784   3.0543   7.2795  -5.1940  12.6105  -9.3077  -2.1609  11.9734
#>   -5.9690  -4.3641  -6.0628   2.1650  15.1780  21.5136  -3.8182   0.9038
#>    1.9888  10.1247   7.6158  -7.5402   8.6767 -12.5719  12.6535   0.8705
#>    5.5595   4.2863   3.5779   1.5700   1.9557 -12.8732  -3.4011  12.1369
#>   -1.6805  -5.7573  -2.7888  17.1723  13.8386   4.6484  -3.6582  -1.4347
#>   -1.5359  -0.2407  -8.2734 -10.3193  -0.2715   0.8504  -7.6677   7.7078
#>    0.5597   4.5146  -6.7693 -11.6341  11.3114   0.1207  11.5439  -1.5291
#>   -5.4278   3.7257  13.1816  -5.4765   4.7410  -7.4410 -12.5533  -4.1024
#>    4.9651   0.2194  12.4301  -2.8568 -15.6471  -6.3487  -0.6189  -1.7127
#>   -4.6419   0.8046   1.7169   4.2360   4.6081   2.6448  17.6810  -3.8485
#>   -6.5498  -1.9627  -5.9084  -6.6436   1.3453  -0.1741   5.0210   4.5610
#>    1.7937  -3.1935   7.9798   9.2398   5.5095 -10.5603  -3.8274  -3.6996
#>   -0.7329   2.5824   3.7840 -12.0182  -2.2886  -0.1516 -11.8617  -2.1732
#>    4.1030   0.7113   4.8599  14.7174   7.1045   4.3235  -6.9919  -7.2805
#>   -3.9277   3.6921   0.0211  -1.7522  16.4537  11.8797  -3.7939  -0.4908
#>   -3.0056  -7.3489   2.8292  -4.0700  -9.8700   3.1231  -4.1670   3.2311
#>    0.5519  -2.8819  -5.9554   6.2930   7.9576  -4.4507   0.3681   3.7784
#>   -4.3144  -1.1849   2.1821  14.3599   0.5177   3.7138   5.0842  -3.5730
#>    0.9524  -3.6497   2.2839  16.7793   6.0761  -4.5534 -19.2376  -4.8636
#>    3.4579  -2.4513  -3.7111   5.4675  -5.9853 -11.9899   5.1590  -1.4773
#>   -0.7849   2.5223   6.5569  -0.0943 -14.5392   4.9234   3.8501   3.4630
#>    4.9113   2.9528  11.5161 -13.2131  16.0984  -8.5377   4.0133  -4.7424
#>    1.3375   4.8425  19.2457   2.2589  13.3208   6.1825   3.5190  -4.3956
#>   -1.9924  -2.3752   8.5795  -4.3589  -0.6156  -2.9404 -12.5247  12.3980
#>    8.0771  -6.9114   2.6062  -6.4136  -6.4995   3.1320   3.0044   1.0694
#>   -1.2472  -6.3712   6.5072  -5.1903  -5.2607   7.0450  -3.4828  -0.8156
#>    7.2004   0.1002  -1.2064  15.6107  11.4599   6.4091  -1.8206 -18.1269
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]