Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.2984   7.8456   5.0861  -3.0145  -0.0495  -0.0088  32.2695 -15.1239
#>   -4.0092  10.1777  -7.9676 -13.8561  -4.0475  13.8396  10.3163   5.3583
#>    4.5541   8.0815   2.2180   5.0493 -17.4190  16.6444  -7.0064 -11.4717
#>   -2.0743   1.7957  -2.7240 -11.2832   8.1513  -4.5136  -3.4309  -0.9950
#>    2.9502  -2.9403  15.3078  -3.5122  10.2548  -6.5312  -2.8286   7.8872
#>    6.3181  -1.5960   1.8643 -18.1159   3.5167  -8.8285   0.1567   0.7405
#>   -1.1736  -2.4232  -2.1592   1.1049   0.2758   5.0118   7.5979  -9.1665
#>  -10.3116  -2.3559 -14.6623  31.5542  -0.2713  -3.5265  -3.0990   6.9598
#>    4.7128   7.9690  -8.9478 -20.1564   2.3185  -0.6317  -4.0680  -0.0831
#>   -0.2603  -3.9583  -5.6109  -1.9323  -9.6324  -1.4487  -4.6064   0.9795
#>   -4.2909   7.5107  16.6043  13.0793 -10.8908  -0.9873  -5.6546  -9.6294
#>   -0.1959  -4.9362   7.2143  10.6208   0.6044  -4.6966   4.5215   9.2867
#>   -3.5214  -1.2374  -3.1605 -13.8291   8.7020  -0.8364   0.1153  -0.9431
#>   -0.4722  -5.6048  -4.6313  -9.6530   8.0569  16.4303  -6.5528  -6.5859
#>   -9.9842   1.4941   0.0546   1.6491   3.4449  -3.4497  -7.8123   3.9197
#>    4.3287  -1.5134  -6.8008  -2.8413  16.7252  -5.1910  12.9473   9.3803
#>   -3.6377  -6.2757  -5.2190  -0.7054  -0.3834  -2.6170   4.4595   3.4238
#>    3.7028   1.3688   4.1654   9.3325   7.7250  -2.3182   6.4751  -5.3483
#>    3.8699   1.1857  -7.1578 -14.6810   8.1843  -5.9652  -2.6193  11.2480
#>   -3.9188   5.3557   7.6369  -3.9327   8.8123   9.0834  -8.9094   2.9407
#>   -5.0521   3.7425   9.0948  -9.6481 -10.1270   2.9039   3.0710   7.5257
#>   -2.7928  -0.4423   4.8126   4.4702  -8.3585   4.5304  14.3479  -6.3153
#>   -2.7454   8.4581  -4.6774  14.3923 -12.3377  10.7240   5.4515   5.8093
#>    2.9071   5.6942 -11.2304 -10.8856  -5.6121  13.4148  -7.6640  10.7911
#>    0.6010  -8.4877  -4.1773  30.7905  -6.5251   3.0520  10.1211   0.3686
#>   -3.0383  -2.5196  -5.8768  13.6513  -2.9062 -16.5825   8.9842  -5.2572
#>   -1.1447   2.2613 -12.9115  -4.7597  -3.6631  14.7656 -17.9742  -1.3843
#>    6.1818   8.6711   4.0971 -14.3223  -4.5230  12.0134  -1.8996  -0.2898
#>    0.0252  -5.0499  10.8645  -4.8575  -3.3848  10.5776   8.4112  -0.7805
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]