Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.0608 -11.1634  -6.4037  18.4135  15.2890  -5.0030 -16.1740   1.3231
#>    4.0251   7.5499  -5.2279 -10.9992  -0.8628 -14.8777   0.1329  -5.9685
#>   -3.4558  -1.5011  10.1042   1.5160  -1.6637  -2.2177   0.8813 -21.9265
#>   -6.3019   0.9826  -5.5303   1.9861  -4.3429  14.0216  -1.9770 -14.0483
#>   -0.9112  -9.5705  13.0056   0.3385  -3.7492  -7.1555  10.6550 -12.0337
#>   -1.8668  -6.7564  -5.2767   3.0799  -1.1378   9.0775   5.8912  -5.1337
#>   -6.5885  -6.8955   9.1826  -0.7270  16.5069  26.4274   3.1113 -18.4895
#>    0.9385  -7.0734  -7.7568  -5.8972  -3.8742  18.2491  19.7124  15.9762
#>   -2.4215   2.6691  -2.1355 -12.9465  16.5958  12.7322  -1.0329   8.1404
#>    0.9049   1.3963   0.7555   0.9947  -6.3494   3.3880 -20.1945  10.4113
#>    3.4126  -2.8621   1.9159   3.7254   3.2793   5.1460   5.8592   9.8121
#>   -7.1069   1.4963  10.1063  -9.6104  -8.2293  -8.0715  -4.0495  16.4187
#>   -4.4849  -5.2107   0.5774   8.5701 -12.3605   2.9142  -3.2227  13.0841
#>   -4.7790  -6.6857 -10.1119  -2.3920  -1.8911  14.0864  10.5448   0.6591
#>   -4.0169  -3.4675  20.9520   1.2804  -4.5075  -4.5569  18.3610 -17.7270
#>   -1.5556   0.4055  -6.2699  -1.5535  15.5710  -7.8329  -7.3405  -0.3679
#>    2.2325  -2.9290   5.8278   3.9484  -7.0227 -10.6919   0.2530  -5.5872
#>    0.2863   7.2851  -5.9877   3.0757  -9.2367 -10.7398   0.4000  -0.7935
#>   -5.4200  -4.0373   7.4263  -9.2199  14.7017   8.8613  11.0232   2.4321
#>   -3.6837   4.7862  -3.7207  12.4087   2.0633 -10.5832 -13.8096   0.5561
#>   -4.8189   3.0111   9.9986  14.8335  -7.3172   7.4855   2.1203  -1.6385
#>    4.3549  -5.3217   6.9993   7.2530 -14.6515  -9.8403   0.5692  -2.2212
#>    1.7388  -3.9636   6.7692   4.9917   0.4147  -2.2516  -3.4050 -16.8509
#>   -0.8211  -0.3107   1.3941   4.8587  -0.3860   4.8258  11.3916   9.8056
#>   -9.1680  11.7702  -1.4836   3.6625  -0.0681   5.3085 -13.8453   6.6931
#>   -2.1747  -2.4455   6.6544  -7.9190  -5.9720   9.4647  -4.4113 -20.2676
#>   -0.1417   5.2045   4.4063 -17.7611   7.8348 -15.8177 -16.4543   1.1643
#>   -4.6408  -1.5309  -4.8875   8.9642  -5.0698   2.6185 -11.1299  11.5574
#>    2.2061  -8.5896 -10.7596   7.9435   1.3984  -6.9616  -3.5740  10.7662
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]