Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.4217   4.1007   4.5545   2.3200   7.1046   5.3645  -5.1045   6.2642
#>    4.7236  13.0579  -1.1490   1.0675  -8.3878  -3.4849   0.0437   4.7102
#>   -3.9612   1.9716  11.0420   1.7105  19.7882  -4.8095   4.4230   2.9582
#>   -3.7713  -1.3823   0.5227   9.7838   2.0055  13.9175   3.2905   9.8158
#>   -2.3205  -9.6363  -0.5445  10.3469   3.1657   3.3161 -11.6031 -10.2986
#>   -0.2038   2.2579   2.6259   6.2223   4.5698   0.6143   2.5923  -8.2474
#>    2.6122   2.5829   0.0510  -8.5879   1.5674  -2.3763  -0.7971  -0.9416
#>   -4.3619   4.9158  -2.7420   3.2085  -2.0072  -1.5638 -14.0185  -0.2441
#>   -3.0246   5.7606  11.3125   2.7945  -9.1881  -8.9678  -0.5690  25.6901
#>    5.5672   7.7696   6.0759   2.3579   3.4868  10.7411   1.2398   2.6533
#>    0.2629  -0.9816   2.9609   8.2904  -8.5746  14.1791   2.2677   3.1097
#>   -0.7822  -7.4235   6.6410  -4.7513  12.0030  13.0615 -10.2563  -7.4280
#>   -3.4876  -8.1381  -3.2504  -6.8084  -6.1049  -1.2307  -0.1750  -0.1155
#>    0.8554  -1.4307  -7.3374   4.8762  -0.4492   3.7649  -2.8791  14.5158
#>   -2.1250   1.9051  -1.7271  -0.7844  12.1817  10.6491   1.9710  -5.6819
#>   -1.3066   1.5243  16.4757   0.8426 -17.5337   0.2687  -4.3946 -16.4418
#>   -0.4547  -2.9727   2.7218  -3.1368   1.8419  -0.3574  10.1181   3.2944
#>    1.3114  -2.1073   1.1065 -14.3121 -17.5622  -5.5733   0.6965  13.0638
#>    3.7166   1.5443  -3.0780  -4.0933  -8.5411  -2.3647  -9.6572  -0.4309
#>    1.3062   3.4605   2.3267  -1.6979 -13.5369   1.1589 -10.6333 -15.4740
#>   -1.5170  -0.1417 -12.9249   2.0407  17.8071   9.3573   8.9200 -17.2667
#>    0.8782  -4.1044  -3.8712 -13.6694  -9.2064  12.0058   3.3352  -8.0558
#>    3.5132   2.0562  -6.2746   6.2755   4.8446 -11.6098  -3.3142  -4.6045
#>    0.9446  -8.5833   1.8682  -1.6114   6.1865   5.0695 -11.5348   0.3955
#>    1.5556  -2.6003   0.0371  -8.9114 -14.7666   7.7007 -16.1803   4.7204
#>   -1.9840   3.0705   4.5695 -10.3467  -2.9708 -18.1178  -5.9171   6.4623
#>    2.3764  -5.0537 -10.7229 -10.5593  -4.5133   0.5989   4.1151  -9.9204
#>   -2.7145  -2.0912  -8.4990  -0.9044   2.6221   3.7623 -16.3742  10.1035
#>   -2.7441   0.1233   4.2444 -18.3806  12.3945   2.2476 -10.4835 -18.2950
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]