Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8 -12.1997   6.4434 -11.7680  -8.1763  -4.5915   5.2810 -18.8939  -9.0865
#>   -1.8019   1.9540  -7.9319  -1.0493 -14.4021  -0.3443   0.0294  -2.3589
#>    4.2839  -5.2405 -11.0537  11.4839 -13.5674   1.9177  -4.5245   5.0190
#>   -7.9092  -1.0587   4.8162  -1.9392   5.2336 -14.4227   3.9313  -1.0090
#>   -4.9951  -2.2374  -2.4701   9.5877  -5.5296  10.3540  -6.3150   0.2163
#>    1.7358  -9.6827   3.7502   6.8107 -11.9167  -0.4959  -6.8123   8.8270
#>    4.0116   0.5399   4.9454  -7.5218  -1.1913 -16.9592  -2.0878   0.1804
#>   -5.6021  -3.1628  -5.5455   4.6283  -8.4472   8.4933 -15.1136  15.2478
#>   -1.3656  -0.7551  -2.4754   3.7076   8.1224   9.0405  -9.4945   9.5489
#>    1.4975  -6.2956   6.0837   2.7843 -14.8431   5.1562 -11.1374  -7.6024
#>    0.1004   4.1091   3.5356   1.3053   0.5580  -9.8194  -4.9350  -6.8364
#>   -1.2019   2.3378  19.1575 -10.7188  29.5487  -8.3661  21.2425  11.3632
#>    1.4494  -6.7317   0.5384  11.0539 -17.9616  18.8689   1.5737 -11.9754
#>    6.5700  -5.5584   5.7578 -13.3752  -0.4278 -12.9883   0.0567   5.8579
#>    1.6287  -7.8567   5.4723   2.6175  -7.1857   0.4531  -0.6964  11.4243
#>    1.1686  -2.9507   6.4185  -5.3074   3.2588  -4.0088 -10.9832   2.5146
#>   -5.4643   2.4911   4.6579  -1.0427  -4.6420   4.9925   2.0650  -5.7845
#>    0.5950  -9.3653   5.1954  13.2261  -7.9612  14.4071   0.0576  15.8660
#>    0.3686   3.8622  15.9373   1.3583   0.7226  -0.8143   4.9129   4.5879
#>    5.1737  -8.6533  10.8378  -2.5213  -7.6865  11.1262  15.0366 -12.8560
#>    0.4880  -0.7596   1.2626  15.7714   2.1057  -7.3631  -2.7173  -3.1069
#>    0.1120   1.8848  -0.6815  -9.3259  -0.3748  -2.0228  -5.4735   3.0430
#>   -9.1036  -5.4132 -10.4005  14.8949  -0.4730   5.0763   4.1977   6.6269
#>    0.6586  -2.8675  -8.1697   3.7585  -2.2848  -9.2810  -5.6827  14.0811
#>    5.0057  -6.2386  -1.0201   3.3305 -21.4769  19.0862  -9.1795   3.3597
#>   -5.3904   1.7429   6.0854 -10.2566   6.4036   1.5230  13.7976  -1.0799
#>   -7.1273   1.6143  -5.1660 -17.7422  -0.1957 -10.5861 -20.5580   0.8712
#>   -6.5445   1.3460  -0.9614  -0.9033 -17.0333   4.3610 -14.4859  -4.3490
#>    0.9713   1.4715  -0.2003  -2.2685  -4.8737  -8.3870  -0.1504  -8.9799
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]