Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.2837  -1.5840  -8.8754  -7.4186  11.8435  -2.9829  -2.7291   1.6572
#>   -0.2332  -5.2037   5.2569   2.5409   4.6349  10.1466  -4.2292  -3.6370
#>   -3.7816  -1.2106   4.0270   4.5395  14.3318 -15.1722 -10.2984  -6.2510
#>   -1.3553   0.1201  -0.0325  -2.7617  14.1406   2.2740  14.8239 -17.6845
#>   -4.4166   5.0978   7.7606   4.4298  -2.1777  -2.4531  -8.1053   0.1006
#>   -1.8707  -5.6814  -5.4327   7.4464   1.8776   7.2796  -6.6291   5.3834
#>   -4.7835   5.4735 -11.6049   2.6627  -2.9674   7.3061 -12.8654  19.6132
#>   -2.9426  -5.5350   4.4043   6.8760   5.1262  -1.7277  -5.2745  -3.7619
#>    1.6540   0.2139   1.7231   3.5747   0.6514   5.7418  18.3570  11.2966
#>   -2.5544  -0.0766  12.9591   6.6112  -6.9697   2.9591  -3.8085 -16.3972
#>    5.8246  -4.4829   3.0143   6.1457   6.2692  -4.2121  -4.2959 -11.6440
#>    1.6598   3.2359  14.1072  -4.9221   4.3410   9.6773  11.0348  -1.0398
#>   -2.4205 -11.6072  -2.5449  -8.5733  -2.7428  -8.8517  12.9019   2.8838
#>    1.1163   6.5698   1.2952  -4.2225  -8.3394  -2.2741  -4.7832  -1.2380
#>    2.0573   2.2579   3.2565  11.2173  11.0903  15.8326  -0.8659   1.3286
#>   -0.6352   2.4408   0.3697  -4.1729  -1.8815  -2.0179   6.2379   0.8627
#>    1.8575   0.3118  11.8955   0.4680   3.3233   8.0515  -3.3767 -13.4063
#>    2.5768   2.7382   1.3353  -3.1651   9.9329  -6.3728   2.8874  -7.3012
#>    1.2881   8.5085  -1.6738   5.8429  -5.6766  -5.1511 -19.0191  -2.0992
#>   -1.0636  -8.1105   3.5191   0.9035  -0.4221  10.0472  -0.7416  -5.5363
#>   -0.6701   0.1339  -4.4042  -6.0616  -0.6248   2.3486  -3.5508  -9.2006
#>    2.9408  -1.8978  -3.3114   1.2486  -8.4182  -0.0123   2.9778   8.1899
#>    5.2539  -1.6929  -7.0324   4.0200   1.5516  -4.6493  -1.9193   9.8558
#>    2.7677  -4.8917  -3.0331   1.7831   6.2028   3.6157  11.5790  -0.8628
#>   -4.6341  -3.1566  -3.6156  -2.3366  -3.0315   7.0922   1.2226  -0.7405
#>   -5.6885   0.7136  -3.3673  -3.9377 -11.8039   0.7307 -10.3225   0.0851
#>    5.8854   3.1628   7.6927  10.4315  10.2641   4.6891 -18.6047  -3.2520
#>    0.0065  -3.7367   7.5100  -3.5916  -0.4142  -6.2066 -10.0641  -2.6437
#>    0.0456  -1.7157  -4.9778  -6.8468  -0.4608   7.0298  -4.3098 -10.0992
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]