Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.2899   0.7760   8.4760  -2.7132  13.6522   1.7970 -21.8283 -22.0079
#>   -5.1525   3.4354   0.4537  -0.7905   9.1742  13.1779  -2.8202   8.7739
#>   -6.2375  -3.3576  -4.1464   1.4187   0.5048  12.6959  -0.6180   9.4879
#>   -3.5828   2.8474  -2.1819  -1.2998 -11.7865  -0.1977   9.4859  14.1933
#>    1.5897  -1.6931  -4.4048   0.1192  -1.9098  -6.7420  -5.2633   6.4606
#>   -6.8006   7.3319  -3.7107  -6.2373   3.5668   5.5690  -8.9346  -9.3716
#>   -0.7423  -3.3455  11.4427  -2.9245   0.2422 -10.8652  -9.5265  -3.6128
#>    0.7364   4.5186 -10.7879  -0.8173 -11.1831  -6.4349  -9.6632  -2.3394
#>    2.4167  -3.5801  -5.2517   3.7087   1.3140  -4.4278  12.7840  -3.6591
#>   -2.2645  -1.0832  -3.7132  -2.5370 -10.3253  -9.6628  -7.2310 -10.0643
#>    3.0449  -0.6824 -13.4779   0.5396   2.6121   7.1123  -4.7051 -10.1635
#>    4.4380   1.0274   1.3939   5.1184 -11.3026 -12.6806   0.0923  -4.6216
#>   -1.3843  -7.0376   4.4459  -2.0627  -0.5241  -2.6011  -2.9461  -8.2706
#>   -0.7671  -6.3910  -5.9440   2.6998 -13.2771   5.9500  -2.6533  -8.3289
#>    2.3951   7.1069  -5.3707  -9.4818   4.0598 -12.0432  -1.4529  -3.5502
#>    3.3615  -8.4948   6.9724  -1.9863  -9.1312 -14.2784   2.2557   1.7094
#>   -0.5203  -4.7079   0.3021  -3.2251   8.2498  13.9010  -0.2143  -4.6921
#>    0.8975   3.3516  -4.1992   8.7710   6.0126  22.5573   0.4715   4.8327
#>    3.2261  -3.1604   7.3726  -2.6246  12.0978  11.2999   6.9430  -1.0196
#>    0.9207  -4.6527  -8.7726  -6.6413   1.5777   4.4578   9.9648   5.9960
#>    0.5972  11.5251  10.0672  -1.5418  -6.4175   4.0825   4.4631  -8.5768
#>    0.8947   3.1823  -7.4628   3.7655 -12.1108  -5.4491  13.4186   3.2246
#>   -4.1539  -0.5514  -3.4485  -7.4875  -2.5152   4.0746   4.4526   4.2412
#>    7.2318  -8.2577  14.1540   2.7452   2.6909  -1.4144   1.7539  12.0812
#>   -3.3157   3.1915 -10.4802  -1.4739 -19.8619  -2.1139  -2.3318  14.8736
#>   -3.8075   1.2575   0.3453   2.3124 -12.7860 -13.6582  13.2796  -2.1601
#>    5.7186   4.0980   2.5750   3.4238   1.3224  -5.6379  11.1776  10.7264
#>   -0.8775   6.0565   6.8863   4.6845  -0.8641  -2.3441  -0.0355  19.9157
#>   -0.6263   6.1494   0.2999 -12.9941  13.3321  -7.7312  -4.8148 -12.3767
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]