Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  7.2545e-01  2.4551e+00  1.0094e+01  1.3245e+01  1.6963e+00  3.0269e+00
#>   1.3622e+00 -7.1677e+00 -7.4107e-01  6.2797e+00  4.0637e+00  3.7354e+00
#>  -5.8513e+00 -6.4299e+00  9.2688e+00 -5.3486e+00 -7.6525e+00 -7.6691e+00
#>  -7.5014e+00  8.0993e+00  4.3131e+00  4.3889e+00 -1.9692e+01 -1.4156e+01
#>   3.2836e-01 -7.5375e+00 -2.4439e+00 -4.6631e+00 -1.0723e+01  5.2193e+00
#>   9.0406e-01 -8.7702e+00  8.5913e+00  7.3379e+00  3.6168e+00  2.9075e+00
#>   3.3588e+00  8.7499e+00  8.0595e-01  7.3075e+00  1.6289e+00  1.1785e+01
#>   1.4272e+01 -1.8424e+00  5.4798e-01 -8.7584e-01  1.0357e+01  1.2193e+01
#>  -1.1724e+00  4.2621e+00 -1.8631e+00  5.5048e+00 -3.8396e+00  2.3582e-01
#>  -5.0653e-01 -2.9783e+00  1.4549e+00 -1.2778e+01  4.7606e+00  6.6118e+00
#>   1.1402e+01  2.6833e+00  1.0474e+01  1.3000e+01  5.6286e+00 -1.3535e+01
#>   7.2357e-01  1.4427e+01  6.6188e-01  2.6226e+00 -6.0192e+00  1.5607e+01
#>  -5.8492e+00 -1.0351e+01  7.1001e+00 -4.6264e+00 -8.0481e-03  8.3075e-01
#>   6.3418e-01  3.6651e+00  1.2014e+00 -3.5790e+00 -4.8769e+00 -3.4987e+00
#>   2.7462e+00  6.1197e+00 -8.6661e+00 -1.1789e+01  5.5804e+00 -9.1892e+00
#>   1.8267e+00 -6.5967e-01  4.2546e+00  6.1629e+00 -5.0242e-01  1.9395e+00
#>  -1.6786e+00  3.6584e+00  6.0423e-01 -9.7366e-01  9.0959e+00 -6.8753e-01
#>   2.5914e+00  8.4443e-01 -4.2464e-01  2.2889e+00 -5.2586e+00  3.7890e+00
#>  -5.6186e+00  6.7556e+00 -7.1823e+00 -3.8032e+00 -2.3240e+01 -1.1491e+01
#>  -5.5664e+00  1.4410e+00  1.1948e+01 -1.0137e+01 -1.8435e+01  1.2446e+00
#>   4.0886e+00 -6.0478e+00  6.1177e+00 -1.8186e-02 -4.0897e+00 -5.0410e+00
#>   4.1381e+00 -4.8841e+00 -4.5869e+00 -3.0007e+00 -6.8413e+00 -1.3747e+01
#>   6.5160e+00  2.3207e+00  5.0794e+00 -1.0637e+01 -8.9426e+00 -4.8262e+00
#>  -7.4221e+00 -1.0116e+00  7.7417e+00  1.0105e+01 -1.0681e+00  1.1910e+01
#>  -2.7248e+00 -5.9688e+00 -5.9239e-01  2.3713e+00  3.6379e+00 -4.1827e+00
#>   6.0847e+00 -8.6246e+00  1.0705e+01 -8.1970e+00  6.5141e+00  2.2023e+00
#>   5.3579e+00  3.0864e+00 -7.6936e+00 -8.7318e-01  1.4517e+01  8.1545e+00
#>   5.0815e+00 -7.0728e-01 -2.5772e+00 -2.4886e+00  8.9330e-01  2.2015e+00
#>   3.8714e+00  2.4279e+00  3.1956e+00  3.3079e+00 -1.2109e+01  7.0159e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]