Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.3007  -1.1931   3.8928   5.9081   1.2951   4.7540  -5.1920   2.4174
#>   -2.4724  -5.5633  -1.8089  -5.0710  -8.2488  -7.9927   7.0987   0.1492
#>   -4.2962  -9.1045  -5.4952  -7.0143  11.7521   4.1465 -10.7421  -9.5644
#>    1.6250   9.0431  -2.3175  14.2066   0.3061  -7.2676   5.2655   4.1566
#>    6.4489  10.5897 -15.4282  -7.1545   1.8176   7.9977  -2.8998  -6.9649
#>    0.6156  -3.4509  -4.0442   8.7039   2.9870  -2.5329  -3.0712   0.1075
#>   -4.1134  -8.6817   0.4037   1.2862   4.8834   7.1293 -12.9701  -3.5699
#>    2.0846  -2.8872  -1.8481   0.3534   4.2788 -17.0831  -2.5758   3.0261
#>   -2.4879  11.3114   1.8138   4.4863   1.9965   1.2308  -6.9652 -12.8120
#>   -1.3364  -2.8564 -13.7958  -7.4754 -12.7830   8.5076  -2.0575  -4.7481
#>    1.5216  -4.0462 -16.9625  -7.9792   4.3302  -8.2856  -8.9206   8.2754
#>   -4.8868  -0.7915   9.3093   8.0335  10.0057   7.1803  -4.9382   2.5675
#>   -3.6025  -2.7001  -2.3527  -3.7368  -3.6852   3.4695 -10.7382 -15.3814
#>    0.6777  12.8122   8.6651   5.7471   7.3113  -1.9092  -5.7134  -6.1314
#>    6.7863   2.6234   5.9346  -2.4326   9.3018  -8.0799  -0.3759  -5.5357
#>   -2.3939   0.9878  -4.7924   9.9153  -6.0933  -7.1879   1.4651  -2.5168
#>    3.7643  -5.6361   8.1369  -7.9198 -16.0700  12.6981   2.8331  -6.1518
#>    0.3841  -6.2421   0.5406   9.0201  10.4767  14.6050  -8.8249  -1.8242
#>    8.7152   8.7019   5.2950  10.0105   9.6214   3.9628  -1.5244   1.1409
#>   -5.4931 -10.5057   2.6145  10.6549  11.9565   7.9593   0.8536   0.8748
#>    0.4302   0.8121  -5.6532  -6.2349  -2.1593  -4.0535   6.7767  -2.8039
#>   -4.5723   4.2955  -0.0406  -1.8183 -14.6737   0.6009  -8.0396 -10.8894
#>    1.8195   7.4356   3.9739   0.8134 -11.8291  -3.5145 -14.6296 -15.3720
#>   -0.4736  -3.3810  -6.7338 -10.6533  -1.6274  -1.3680   7.0758  -0.8512
#>    7.2375  -2.9103  -9.9479 -15.6544   1.6103  -2.8882  -9.5491   8.0299
#>    0.2533   0.9914  -4.9229   3.1902 -10.9507  13.3467   4.6874  -6.2331
#>    0.1121  -0.7747 -15.5615 -12.3832   1.1346  -9.4911  -8.7243  -8.7910
#>   -2.6998   4.8796   6.5217   3.5896  -0.1377  -1.4481  11.1686  17.5964
#>   -0.3245   0.6818   9.4615  -4.1082  -3.8368  -7.9500  -3.2605   6.4949
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]