Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.2296   0.1547   1.3082  12.8675  -8.0739   9.7773   9.1990   3.8138
#>    8.4824   2.2725  -5.8971   6.1721 -13.2600  -4.1307  -3.3147   4.9203
#>   -3.4960  -3.2566  12.5323  -1.9416  17.2656   4.6997  -9.8475   3.0074
#>    5.5995   1.9557  -4.6120 -12.6131   9.7370  10.1832  -8.3751   2.1740
#>   -4.5209   2.3918  -1.1015   7.4299 -10.5433   0.0755  10.5033   5.3596
#>   -3.7107  -2.0874   6.1674  -8.3061  -9.8415 -13.9651  -0.5621   6.6164
#>   -3.7570   2.0494  -1.1467   3.5188 -10.0904 -13.7032   2.4617  16.4706
#>    0.9021  -4.9660   3.8873  -5.4561 -10.2393  -7.8425  -9.3319  14.5627
#>   -4.2517  -1.8965  -6.0191   5.1477 -11.6749   6.2882  -0.4098  -4.5378
#>   -5.3016  -2.2304  -7.2770  -3.7447  -0.1918  11.9091   4.7879   9.0239
#>    1.3982   0.8150  -2.6353  -5.0757   5.8087   9.0534   6.8436  -9.2663
#>    0.3063   8.0357   9.4122  13.5813 -12.6023  -2.9786   1.6946  -1.0413
#>    7.9496   5.1616  -8.4647   0.9001  -4.1849   3.7946  -2.6684  -4.4687
#>    0.0660   1.4733 -10.2176   5.0069  10.2184  -6.5741   4.2609  -0.2994
#>    0.0699  -5.8164  -1.4373   1.3195   0.6442   7.3871  -6.9664  -6.2578
#>   -2.7436   1.1928   0.8491  -4.9311  -0.4644   8.5224  13.2251   3.7756
#>    1.2473  10.0562   6.4007  -1.0815  -5.4618  -9.2684   6.2536   3.1739
#>   -0.1982  -4.2691  -7.7968  -3.4452  -8.8813  15.9694   4.8148 -10.6208
#>    0.6356  -2.9413  -1.5191  -8.1063  11.7396   2.3325  -4.5025  -0.0291
#>    6.6715   6.0536   3.3265  -3.8789   9.2747   8.3641 -14.2526 -10.8088
#>   -0.5700  -0.0139   2.6374  -0.6456   2.6451  -4.1326  11.2910  -5.8457
#>    1.8906   5.4596   1.6222 -17.1135   9.4011  -4.5305   7.7356 -19.0698
#>    0.8958  -5.3845   8.1503 -14.0704  -1.6330  -1.0081   0.1949   7.6943
#>    0.9096  -0.0448   4.3962  -2.7957  -9.9969  -9.6237   0.9364 -17.5134
#>    0.4073  -3.0450  -3.7700   2.8659  11.0906  14.0922   3.9037 -10.1763
#>    1.1184  -9.2795   0.7603   5.9543   9.3594   1.0970 -10.7176  14.5505
#>    8.0759  -0.2686  13.8627 -15.8042  -5.7867  -9.0530   1.0454 -17.4633
#>    5.6633  -1.7385   0.1053  -9.9474 -15.2073  12.9636 -17.0971  -7.5567
#>   -9.8304   6.8368 -17.2382  10.8708  10.7863  -7.1593  -3.5978   5.4843
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]