Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   9.1858  -2.7440   7.8398   5.9641   8.3286   7.5210   8.6008   0.3006
#>    2.8443   4.2773   7.2779  -1.6333  -3.7006 -10.6969   6.9409 -10.0193
#>   -1.4722   0.5746  -5.0079   1.2638   7.1499  -3.9985 -10.2452  -3.7025
#>    4.2048  -3.0172   1.3833  -3.1172  15.1044   7.6625  11.0084  -5.3830
#>   -2.9797   7.9726  12.6789  -1.7629   1.4551  20.7202  -0.2590 -15.3886
#>   -4.2641  -4.9081  -2.8412  11.3269   0.9502   0.9441  -8.3749   6.7381
#>   -4.3092   5.8714  -1.5413   7.3438  -8.1618   6.1225  -0.6303  11.9830
#>   -6.9643  -0.3103  -7.1234   7.4437   0.0087  10.3628  -7.0501 -10.8756
#>   -7.6366   2.7305   2.4031  -2.1522  -1.0287   1.8486  -0.1044  -5.8944
#>   -2.1657  -1.1435   9.1315   6.8749  -3.3011   8.2289  -6.6192  -6.2894
#>    3.6759  -8.7952  -7.0160  -6.7388   3.3991  -1.4672  10.3314  12.8621
#>   -0.2528   6.3963 -14.2011   7.0064   7.7812  -6.9807 -14.9260  14.6215
#>   -4.0577   1.3391   0.0362 -10.1584   5.1565   5.9457  -8.4073   9.2771
#>   -2.9313   1.3295  -6.1185   4.0073  15.4664  -2.6706  -6.2419 -15.5923
#>    7.5142  -4.6280  -0.2010  -4.4476  -5.0735  -6.4970  -4.3083  28.8501
#>    1.8346  -3.5550   1.7456  -1.2600  -0.2345   9.1992  -6.9320  17.1286
#>   -0.1912  -5.8840   3.1006  -7.3447   3.5208   1.6654  -8.5470  -2.7493
#>    7.6041  -1.5285  13.4404  13.4557   0.0500  18.1140   9.4177  -2.8154
#>   -0.3072  -3.1937   1.4267   6.3495   0.6238   2.5618  11.6556 -19.3676
#>   -5.5373  -0.9643   1.4208   0.0882   9.1918  -0.5257   2.7586   1.0214
#>    0.0970  -0.7228  -2.6385   0.1972  -5.5341  -4.5720 -17.1845  -3.5216
#>    3.4960   0.9181  -4.0619 -10.7298   3.8036   1.5619  -6.1405   1.2692
#>    1.9043   6.1585   7.1934   5.7434   5.6346  -4.5128 -10.5280  -6.7962
#>    2.9865   7.3978   3.9533 -11.0594   5.0778   8.7958   6.1408  -5.7148
#>    0.4651   0.6437   0.3026  14.3469  -8.3893   6.0066  -9.4659  -3.4672
#>    3.2311  -2.9822  -5.5563   4.0102  13.4564 -11.4480 -10.2981  -5.6830
#>   -2.4693  -2.3219   3.4735   9.1423   1.0840  -0.4416  11.3224  -4.9804
#>    3.3135  -6.1818  -3.2679  -3.3794   9.1902   0.3497  -4.1499  -2.3273
#>    3.5009   1.0676  -4.0889   7.5681  -3.0617  12.4972   3.9387   8.6972
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]