Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.7240  -3.0641   0.0189   0.0472   0.2735  -5.8759  -6.0375   2.5331
#>    3.1909  -8.0738  12.0505  -8.7404   0.0509   5.4375   4.5436   8.0635
#>    9.6188  -9.9918   5.2779   0.5706   4.6724  -4.0056  28.9814   3.9059
#>   -4.3770  -3.3407  -8.3655  -3.9879  -3.5001  18.6159   5.4769  11.5422
#>    3.1708   6.3116  -2.0507   4.7385  11.3236   0.8588   8.0156  16.2265
#>    3.2955  -6.8726  -2.5014  -3.3546  12.9494  16.2484   3.9113  -2.9774
#>    6.1794  -9.1768  -3.2186  -1.1869   6.3772   2.0939  -4.5652  11.4561
#>    2.6656  -7.9031   3.9923  -2.8224   0.4051 -17.0220  25.7683   9.7186
#>    3.5184 -12.6916   4.6766  -2.7659  -4.4859 -21.1607  -6.2009   4.6019
#>   -7.6468   7.0057  -7.0161   0.2369  -7.8271  19.3186  -3.1404   3.0919
#>   -1.2470  -3.9051  -5.3056   5.1965   1.6495  -9.6037   8.8746  12.3276
#>   -5.4821  -1.5161   5.6418   2.6265   5.9555  10.5084  -3.1869 -11.1304
#>   -4.1374 -11.7467   5.5752   0.2006   2.6746   3.4623  10.9565   2.0875
#>   -2.7332  -2.3748   2.5291   6.2106  -2.8997  -7.3426  10.6244   2.7501
#>    3.6458   6.4050  -3.6294  -4.0200  -7.4291  10.4516   8.1476  -8.4415
#>   -8.8169  -0.9433  -8.1033  -9.0225 -17.6020  18.4425   2.0725 -13.2644
#>    7.0213  -0.0336  -3.8730 -11.4385  -4.7419   4.0034  -3.4877  -3.2246
#>    3.7902  -4.1020   1.0272  -8.7782   4.9696   2.9898  -4.9672   1.3046
#>    3.0639  -1.0451  -3.8307  -5.7373  -9.7907   5.4504  -3.1255   1.6995
#>   -0.7652   8.0996 -10.8498   8.2611   2.4318  -2.6849  -4.6482   0.3092
#>   -2.4148   7.6196  -6.5885  -1.5484   4.6128  -7.5071  11.6061  -0.6439
#>    0.7498  -4.5700  -2.6029  -1.9793  10.9144  17.8671  -4.9363 -19.9130
#>    1.1692  -6.6896  -5.4087  -2.1462   0.2488  -5.0110  -0.0676  -5.7294
#>   -2.7438   1.4589   7.6417  -8.9565  -1.9059 -29.8263  -7.9856  -3.4165
#>   -0.3011  -2.5749   3.1004  11.7910  -4.4575  12.9143  28.6247 -21.7956
#>    0.6254   6.9096   0.4656  -0.4572  -1.4428  12.7286  -9.7418   8.4439
#>    2.2544  -7.4770 -11.4156 -14.6107   1.5695  -3.9861 -26.0075   2.9588
#>   -0.7246   4.6888 -10.2532  -1.0899   8.5409  12.8059  -4.1736   8.0009
#>  -11.2638   0.5125 -12.5340 -10.5875  13.7953  12.4451 -12.8504  17.6553
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]