Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  11.0260   3.8981 -10.2888   8.6750  -1.3507   3.3470   9.4347 -19.3714
#>    1.3638  -0.8764   7.6177   4.9407   5.1114   5.8992  -0.7580  -4.0750
#>   -4.7518   7.0962  -2.8591  -4.4815  -8.9989   5.5754  -1.8383  -1.7245
#>   -0.7549  -3.5807  -8.9476   2.6965   2.0347 -13.9478  15.3856   8.1830
#>    6.1322  13.3953  -7.1994 -10.3130  -2.2699  -4.5342  -8.2465  -2.8894
#>  -10.2707   0.4690  16.8597  -3.6330   0.4910   1.1447   4.7415 -17.1261
#>    1.1802   7.6568  -7.7147   9.6647  -2.3798   4.2679   2.4326  15.0069
#>   -3.7753  -4.8804  -0.7026  13.3731  -1.3086   6.6032  10.1538  20.2316
#>   -8.3114  -8.9979  -1.1117  -8.7637   2.9567  -1.4442  -8.9084  -1.1819
#>    4.0085   5.4466   1.2118  -3.7778   4.0167  10.3116  -4.5332 -11.1361
#>    5.8215  -4.4411  -2.3986   5.2605   1.2240   1.1795   8.6614  -1.2160
#>    4.9118   1.0592  -0.9142   6.9168  -5.1285  -4.8159  -2.2642   0.9753
#>    1.1168  -2.0523  11.9827 -14.0774  -7.7069  -3.2237  -2.7249 -12.3957
#>    4.9927 -13.1829  -3.3114  -2.8526   6.1949   7.5924  -5.3419   2.8065
#>   -1.6366  -7.2946   6.3347  -1.1780 -10.4696   2.8425 -10.5409   5.1652
#>    8.1870   2.2869  11.1188  -2.7352 -13.0351  -3.0342  -3.7102 -12.1428
#>   -7.5264  -1.2877  -3.7659  -4.8453   2.8428  13.5774   4.5802   7.5520
#>    4.0558   0.4186   1.8676  -6.0754   6.0379   0.9644  20.1727   1.9682
#>   -1.8961   6.2703  -5.7878   8.8038  -4.1689  -0.2984   6.8743   8.8512
#>    2.5382  -8.3246  -2.2533   2.8191   6.5570  17.1474   0.5644   7.9886
#>   -2.9261  -4.2674   0.2623 -10.0541   8.9750   0.5807   2.2497   3.5663
#>   -7.6807   2.8212  11.4748 -20.4872   3.5867 -13.9811  -2.7742  -2.0191
#>   -3.7391   1.8612  -8.0099  -1.0327   5.5489  -5.2459   0.5466   3.9894
#>   -2.7469   5.4896  -1.3632  -1.8967  10.4218  -0.8749   5.9902   9.6165
#>   -3.8893  -4.5696   7.2938  -5.7745  18.1933   9.0036   1.3580  -7.8281
#>   -3.8317   1.0947  -3.8378   4.5499  -4.9893 -15.0847   0.0534  -2.0666
#>    1.6373  -3.0191  -3.4293  -3.8824   6.1751 -10.9518 -12.6941  -0.7611
#>    0.0097   6.5268  -3.7811  -1.2664  -5.7344 -15.5362  -0.9263  -1.4260
#>    0.3017  -0.1520  -7.7006   3.3755   1.5795  -1.3848   3.4976  -6.7068
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]