Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.4162   3.9385   4.8606   9.4209  10.9901   0.1779  12.6397  -6.2206
#>    1.1985  -1.5907  -4.2320   9.1166  -3.6583   4.6778  20.6747   4.9316
#>   -1.6325   2.6202   0.2442   1.4872 -10.3880 -25.3388  -1.8032   2.2433
#>   -0.2227  -2.5087   0.0546  11.3094   8.2044  -6.9623  -3.4880  -0.9483
#>   -2.5955   0.3204   4.3791  -6.3562   6.1599  -5.2242   2.4731  12.1858
#>   -4.6472  -0.4187   7.8580   5.4269  17.9456   1.1150 -18.7436  -3.6265
#>    1.4410  -2.2755  13.2620  -5.9258 -16.6697   9.0801  -0.1089   5.4994
#>   -2.3341  -0.3791  11.5629  -0.0813   4.2913  -4.1853  -2.9869  -7.7919
#>    1.3585   5.9150   4.0822  -3.3816  -3.7690  -7.1574   3.0545   4.6154
#>    2.8596   1.2022  -4.6429   1.7763  -6.3946 -14.4127   7.1842 -15.5023
#>   -6.0024   3.3161  -0.1110   3.1322   6.4618  -0.3034 -11.3056   2.0871
#>   -0.3943   9.3433  13.9777   1.2035  -2.6829   9.8772  -8.5612  -3.5487
#>   -1.2434   6.7909   6.0524  -0.7055   3.7026  -6.5511  -0.4399  19.1333
#>   -4.5545   0.6676 -12.0240   2.9776   7.4468   1.3210  -9.5530  -8.4309
#>   -0.4124   3.6320  -3.8076   9.7121   0.1901  -6.2284  14.5961   2.0942
#>    0.6645   2.3295  -5.9965   5.6069   3.5362  -7.7518  -0.4229 -19.5383
#>    0.7018   4.9282   1.9123  -8.3351  -1.1287  -0.2105   3.9645   9.8020
#>    3.6993 -10.8710  -5.2910   4.4602   3.8991   5.6585   2.8893 -16.5761
#>   -0.0444  -0.8884   3.2465  -1.1739 -11.6564   3.4618  -9.2710  -6.1818
#>    1.5749   6.7208  -4.9159  -0.2359  13.3085  11.7533   5.1574   9.4406
#>    3.5354   5.7279   0.2230  -1.8271   5.2991   5.1780  -4.3129   4.6480
#>    2.2003  -1.7934  -3.3551  -1.3757   7.7482  13.0612  13.7097   3.8545
#>   -0.2323  -5.9179  -6.0529  -3.7040   6.6653 -15.1893   1.5761  -6.0038
#>    1.9083  -1.4599  -1.1537  -0.1701  -7.1261   5.2947   7.7359   7.0874
#>   -1.2755   0.4305  10.1138   1.4719   1.1118   5.5515 -13.9283 -11.7612
#>   -2.1912   2.0716   4.1777   3.2619   4.5771   0.1791  -3.0212   0.8226
#>   -2.1568  -3.4744   2.2140   1.9934  -7.7821  -3.7273  -8.4033   4.4597
#>    3.0668   3.1488  11.2883 -10.9493   2.4061   0.6983  -2.3130  -0.8691
#>    0.3309  -1.2137  -5.3510   4.8634  -0.3442   0.9277  -2.0724  14.9572
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]