Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.6330   4.7175  -2.6550 -12.0401   2.1328 -15.4282  20.2042 -12.3517
#>   -0.6670  -2.9201   3.3694   4.0228  -1.9284  -2.2517  10.8479  -3.0855
#>    2.2046  -5.4876   3.5916  11.5728   4.4721 -13.3567 -11.2152  -2.9033
#>   -2.7743   9.1725   5.7938   1.8536   2.3349   2.0402   7.5487  -2.9761
#>   -0.9922   6.9180 -12.8218  16.5590  10.0815  -6.0061   1.3118   9.2085
#>    1.8916   1.0215 -17.0084  -9.5777   5.5372   0.8724   9.3440 -12.1306
#>    7.0586  -1.0475   3.9680  -1.0465  -1.8718  -2.9448 -13.4117  -8.7341
#>   -1.2151   2.0846  -3.5818  14.2555  -3.6666  22.6171   6.2944   7.6721
#>    3.6673   1.2917  -3.0970   9.2717   1.6688   2.3289 -10.0991  -7.9989
#>   -2.1545  -0.3307   6.1057  -0.6377   7.2219  -9.2645   3.5851   7.6058
#>    4.0079 -11.6870  -2.6313 -16.1472   2.0727   0.2290   3.9451 -10.5863
#>   -0.3813   1.9745   4.2995  15.2573  -8.8595  -5.8617  -4.5064  -6.5325
#>    0.0290  -5.0594   3.0111   3.9244  13.2932  -6.0245   9.6016   1.2724
#>    7.6137  -8.6666   8.1545  -7.7098  -1.9261   0.1646  15.6475 -11.6333
#>   -1.6307  -4.4769   2.8053  -1.7900  -4.8813   4.8848  -7.1764   4.6284
#>   -0.3678   5.7836   4.2245  -4.8575  -4.5188 -14.5031 -16.0106  -0.7865
#>    0.3648   2.5395  -2.3478   6.8298   0.1875  -0.2428   8.0632  10.1072
#>   -0.2644  -2.7449   5.6517  -6.4328   4.8872  -7.4054  10.5894  -2.4125
#>    0.3043   2.4384  -0.5620   3.4766 -14.3414  -1.9013 -18.1401  -8.2608
#>   -1.9747   1.2266  -6.4953  -6.6209  16.8115  17.2249 -13.4345  21.1657
#>   -4.1259   5.2491   1.5795 -13.3855  -8.9958  16.2991  10.1035 -10.4748
#>   -3.3954  -0.5640 -13.6241  12.7858  17.1600  -5.7994  -5.5198  13.9887
#>    1.9620  -2.0447  -4.0990   3.8841   7.9337  -1.5372 -14.9256   0.1929
#>    0.0798   1.1899  11.4167   6.8863   6.6274  -1.7963   5.5914  -3.7105
#>    0.7453   3.6125   0.0245   0.5088  -6.7614 -12.0528   2.8625  -0.7981
#>    0.6621   2.2064   1.1262 -14.3606  -2.0375  -1.3951   9.5280 -12.5288
#>    2.1081   2.7691   1.0352   5.3535 -20.7666 -10.9472  -6.1455  -5.9195
#>    6.0030  -1.5197  13.7305  -0.4146  -1.8155  -1.5468   2.2035  -0.2806
#>   -0.6755  -4.0971   4.4301 -17.0813   5.5093   4.2021   9.1668  12.7338
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]