Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.8637  -0.2481   5.0034   2.3607   4.4365   1.3799   5.9066 -19.7150
#>   -3.7877   3.8715  -0.3101  -1.5848   2.5063   9.9068  12.9473  16.3746
#>   -0.7763  12.7737 -22.4835   0.2392  -0.6508 -16.9573 -11.7658  24.4972
#>    5.1245  -3.9130   3.5678  15.5006  -2.4338  -5.8014  -9.2623   5.8026
#>   -7.8940   3.7647  -0.2081   3.5594 -12.9979  13.2700   1.9163   3.4232
#>   -0.2100  -4.9453   0.1687   8.9495   1.3717   6.0134   0.2281 -10.9204
#>    3.0426   5.9902   1.9261   1.6747 -14.9899  -4.7898  21.7866  -9.0964
#>    1.2041  -9.3004 -17.2998   2.5125   0.1550   5.5837   4.8100   7.0228
#>   -1.8255   4.7716  -4.1709  -1.4133   7.8305  -7.9207 -13.8563 -13.4060
#>    4.0246   4.0879   3.1638   0.9116  16.4278  -2.3809  -2.3980 -19.3821
#>    4.4778   7.7360  -7.8238  25.7609  -9.9318   9.6905  -3.6173   4.9556
#>    0.8575   3.8368  -5.8126   5.3146  -3.2310  -4.3178   2.4157   2.3819
#>   -1.7210   2.5480   3.8002  -1.2535   9.6418  -0.9243  -1.3497  -0.1431
#>   -5.0638  -5.2613   4.7117  -8.3071   3.5937  -4.6683   8.0891  -2.0890
#>   -5.6522   5.3686  -5.3528  13.8556  10.0494  -3.4554   5.3740  -4.3520
#>   -2.0376  -2.1195  -3.4169   7.8306 -24.2201  -6.9836  -9.2147  15.7791
#>   -5.3222  -6.2357 -13.8138  -6.5503  11.0320  -1.2508   4.4372 -15.3892
#>    2.9314  -3.9298   7.4590   3.1434  11.5615 -14.2917  -2.7611  10.4262
#>   11.0749  -0.4326   8.7122  11.3524  -9.7133   3.5209   9.2609  -4.3498
#>    5.9737  -7.2912  -2.9807   8.3325 -14.5360 -21.8112  21.9633   2.1912
#>   -3.1942  -6.0092  10.9595 -11.3133  11.7752  -2.0796  -2.9222   4.8056
#>   -0.2604   4.9589  -6.0840   2.1653  -6.8801   0.5243   0.7009   4.3445
#>   -7.4249   5.0980   6.8428   9.2400  -8.4868   7.1280   3.8052   8.4812
#>    2.4382  -1.5596  -2.4605   4.0136  -1.3778 -12.5794   1.5558   1.3229
#>    0.3685  -3.0429 -10.3436  10.2729  11.2271   0.8637  -6.6745   2.1024
#>    2.9325  -0.2663   4.0135   0.1583  14.0407  -3.6160 -11.7369   5.0179
#>    0.9728   0.3345   8.3326  10.4889  -8.7997   2.2490   2.7152   0.0943
#>   -1.3921 -18.2145  -6.1436   0.7169   4.1546  11.6827  10.5676  -2.1374
#>    6.7489  -1.1864 -10.4218   7.5411   4.1941 -11.8860   2.4490  16.7250
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]