Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.0885   6.1002  -0.2276  -8.4922  -6.4047  10.4223  -2.2962   1.0150
#>   -0.9572   8.4442  -1.7945  -9.2465   9.0881  -3.0904  -5.7949  -3.5668
#>    1.4568   4.1839  -4.8203   4.3444  -4.1994 -10.7744   5.7496  -0.5058
#>   -0.3397  10.3972  -1.0782  -3.8754  13.5287  18.3622   4.3620  -4.0085
#>    1.5872   2.1192   8.3397   4.8197  -2.8304  14.9068   5.3171   9.1941
#>   -6.7330   1.5030   0.9512   0.2325   5.1419   3.9804  -6.2409 -10.1946
#>    0.8811  -2.4195  -0.2156   4.3844   8.6597  -8.1217  -1.3157   3.8360
#>   -2.3164  -0.8865   4.4311   2.6305  -2.7944   7.1073   9.7466  10.9504
#>   -2.8690   0.8701  -9.4299   0.2025   3.7403   9.5554   0.5443   2.0224
#>    0.7168  -7.0002   8.2805   5.0953 -20.6992   3.6764  -8.9225  -2.2333
#>    1.4700   0.0363  -0.8482  -2.8518  -1.7300 -17.3761  -4.4970 -11.4327
#>    6.5312   5.1113  -3.0881 -10.1988   1.5092 -19.8730  -5.1915  -1.8651
#>    1.8187  -3.0609  -1.4628   3.3578  -0.1675   1.7325   8.4229   1.8282
#>    0.6005   0.6598   0.1667   1.0097  -5.0934   2.3103   5.2800   1.2114
#>    0.7267  -2.7918  12.2456  -4.4412  -2.6984   1.2273   2.7187  -8.5074
#>    4.0993  -8.3846   2.2633  -7.3016   3.4780  -2.3891  -1.0308 -10.2254
#>   -1.5564 -10.2414   1.0359 -19.3072 -14.8970  -4.6362   6.6044 -16.8352
#>   -2.5217  -3.6841   4.2718   8.2437  -3.7817  18.7394   2.3369   4.1077
#>    3.1399   4.7537   4.4898   3.4023   1.6220  -9.1948   6.7721   2.9207
#>    4.5451   2.6836  -9.0240 -12.8264 -10.9090   3.0896  -9.1952   2.1623
#>    0.8646  -2.8753   2.2479   2.0645  -0.8239   6.4499   6.0659   8.4662
#>    0.9370   3.9761   2.2339   3.5399   2.1030  15.1435   8.1942  -0.6979
#>    2.5172  -0.2133  -7.0236  -2.8595  -0.2953   8.9888   2.9481  -3.9900
#>   -0.1575   1.5116  -3.4173  -3.7394   3.4263  -3.8032  -0.8073  12.5521
#>    0.0198   2.4378   8.8777   6.4340  -0.8953  16.6906  13.6225  -5.7903
#>   -0.6953  -2.3980   2.1021   8.6708   1.0490  -0.6426   4.9455  -8.4546
#>    2.6338   4.8335  -1.7223   2.8718   1.1317  12.3888   0.5400  10.7529
#>    2.1578   0.0533 -10.0627  -2.4046   1.0775  -0.2409   2.1479  14.6004
#>   -4.2997  -0.7697 -10.5037  -2.8980   9.6144 -11.1577   4.1218  -1.0829
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]