Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.1763  -7.0707   0.1337   5.0048   2.4511  -4.2012  -4.5185   6.9463
#>   -0.0543   0.2160  -7.4240   4.8299   6.5171  -3.4719   8.6207  -8.7844
#>   -0.9412  -3.9678   3.6352  -7.9480  -0.7841   5.7271  -2.6916  -6.3473
#>   -5.5157  -0.1613   4.1175   8.0380   4.7042   1.6187   3.7513  -0.2047
#>    5.4398  -7.4629  -8.2820  -0.5466  -0.7519  -6.8655   9.6733   0.8936
#>   -1.0225  -7.4334   3.6575  -1.8310  -2.0258   9.9849  14.6893  11.0963
#>   -1.8174  10.0174   4.1749 -16.6304  -3.0502  10.1651  -2.5543  -4.1587
#>   -2.5239  -8.4019  -2.7791   4.7305  -4.0273  -7.0983   9.8734  -7.6202
#>   -0.8746   3.6059  -2.7773  -4.7621  -2.6522  -0.6371   4.2820  -3.1154
#>    2.9623   0.8769  -4.8461   0.4892 -12.5224   2.4513   1.9361   6.5404
#>    0.1749  -0.6922  -1.5141 -15.6314  -7.9961  14.1276  -6.7651  -5.8055
#>   -2.8806   0.2505   3.8352   2.2817  -9.8710  15.4191  -4.0603  10.3634
#>   -0.6764  -2.1818  -5.3152  -7.7647 -12.9728  13.5326   5.6589  15.5180
#>   -0.4172  -4.9078  -0.2036   8.8026 -17.9463  18.6266  -2.5773  -8.3780
#>    0.8070  -0.2057   1.5735  -2.6880   7.7155  11.5801  -8.2499 -13.8145
#>    3.8210   7.7084   4.2764  -1.7090  -5.2335  -0.4029  20.3689   0.6226
#>    0.6736  -9.5387   1.8853  -2.8513  -0.7266  12.3980  -8.9757  18.1765
#>   -2.6206   2.6416   2.6188  -1.5950   1.3079   4.2908  -5.3178  -1.6447
#>   -1.8923  -1.5103   4.8995   3.8792  -0.9640  11.0615  11.7038  -3.0125
#>    1.1528  -4.5278   6.2048   2.3772   1.7861  -2.2540  10.8556  -4.0629
#>    0.1836  -1.3929  -0.3402  -1.7364   9.9619  17.4386   4.4320  12.4850
#>   -2.2081  -5.3179   4.0500   8.0615  10.1397   0.0831   0.8381 -12.0971
#>   -1.7739   6.1694  -2.3946   6.5363  11.6574  -9.3011  -4.8338  -1.2926
#>    1.0802  -5.4375  -6.2436  -6.7120   7.3493   7.6561   2.9943   5.2581
#>    4.7396   0.0910   0.7833   6.4421   9.7513 -11.5572  20.1164  11.3076
#>    3.0466  -3.2046  -1.2766  -6.0495   7.5398 -17.2415  -9.7916  -6.5161
#>   -4.2214   3.2995   1.7887  11.4517 -17.0338   0.9654  -4.3892  -4.8266
#>   -1.5615   7.9008   2.8480  -6.1696   4.6191  -2.8635  13.3934  -2.4427
#>    2.6948   2.6818   5.6965  -3.0287 -15.0732  10.0308   3.5000  12.0362
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]