Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.5325   0.6168   8.0253  15.6225   6.6627  -6.0580  -7.8327   2.9326
#>    0.9729   2.1453  -4.1174   4.7552  -3.1423  13.5297  -5.3600  -9.9586
#>    3.7190  -5.7596   0.4148 -10.1232  -0.2114  -8.8874  13.8947 -29.7703
#>    2.0127   4.9825   9.1844 -14.9575   0.6242  19.0626   6.9502 -14.2982
#>    1.5540  -1.7832   5.1873   4.4024  -8.1002  10.3843  -1.8356   2.0592
#>    4.6151  -6.1721   0.4193  14.5989  -3.7804  -2.3922   7.8191   8.0265
#>   -1.7599   0.7405  -3.7374  -6.8741  -2.1852   2.3982   4.6889  -2.5851
#>   -3.3029   3.6310   3.1051   8.6730  -7.0918   4.8392  -5.7935  -4.4536
#>    1.4083   7.7740 -10.4933  -4.6200   0.2785   5.6202  -0.5737   3.6544
#>    3.2791   6.0233   1.6814 -14.7333  -1.3097  -6.9230  11.7859  -6.1505
#>   -0.3328   3.1359  -8.7587   2.4681  -6.2758   9.0272   8.4501  -2.6630
#>   -2.3658  -3.0251  -4.6121  -2.6917   1.2772   1.2451   4.5666   7.1314
#>   -2.0916  -5.4595   3.2922   1.1273  -5.5625  -0.6340   1.8267   9.6048
#>   -5.7695   2.8837  10.7670 -11.9843   1.4937   7.1737   9.7138  -7.0253
#>   -1.8254   1.6123   5.1112   2.4403 -12.1068 -17.7704  12.2969   4.5077
#>    4.1588  -2.9520  -2.1801   1.1286  15.7386   9.7276  -9.7943   6.5496
#>    4.4401  12.7612  -6.7850  -6.4860  -2.1873  13.4621  -1.5788  22.8797
#>   -4.8767  -2.4229  -8.6272 -10.2680  14.6064  -9.4624  -5.8832 -19.7870
#>    2.0258  -3.4482   5.8916 -11.9127   4.6567  -7.4149   4.0071  -2.6064
#>   -0.1392   3.6189 -11.2965   5.8196  11.0841  -3.7309   1.9738  -4.4466
#>   -2.7291   6.8406   3.1109 -10.4744   4.1179  -4.1368  10.3877  10.0996
#>   -3.6393  -1.3543   0.7682   9.6278  -1.5154   0.2563  -3.8088  -0.6108
#>   -3.2325  -7.1424  -2.7327  -2.0965 -14.8950  -2.8065  -0.1813  -1.4872
#>    2.5195   3.2961  -2.9850   7.0731  12.4441 -12.4411  -6.5738   8.5615
#>   -0.1984  -7.3798  -2.4302  10.4979   1.5952  -5.2145   2.5297   5.5000
#>   -3.3804  -1.9127   4.2572  10.0552  14.3935  -0.5818 -12.2009 -12.7419
#>    4.2522   5.6105  -4.7583   1.8247   1.9625   3.0311   4.5676   7.3275
#>   -5.0131  -4.5892  -4.2195  -1.9320  -8.0736   0.1796  -7.6498  -7.1524
#>    1.3351   7.7539   4.4159  -8.1991   1.3508   3.9473   5.1946   4.7796
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]