Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   7.3687   7.1823   3.0269   3.0148  -4.5972  -6.1147  14.6853  -2.8386
#>   -2.0902   0.5286 -19.8242  -2.9520 -14.4092  -6.4879   0.9157 -10.7223
#>   -8.4517   0.7170   9.0922  -0.3494  -3.7125  -5.5196   4.2792   8.7082
#>    0.6067  -1.0198  -6.9136  10.4432 -14.1408  -1.6693  20.0898 -12.0779
#>    9.1985  -5.4704 -11.4124  17.1742  22.4198   4.4524   1.8406   7.4756
#>    5.7174  -6.1928   1.9457  13.5967  -6.6084  -1.4243  -8.5358  -1.8944
#>   -5.4695  -5.8033  -6.4925  -4.1988   2.9546  -2.5545  -0.2460  12.5398
#>    4.6887  -3.4894 -15.0877  -5.3613 -17.6194  -8.8205  -5.5861  -7.7461
#>    0.9536   0.0670   5.6249  -2.1580  11.6428  14.7547  -2.1114  -4.2893
#>    3.0215   4.6470 -10.2192  10.0802  13.7961  -3.0002  -4.2069  -6.1731
#>    4.9630  -1.3514   4.5283 -15.7619  12.1978   3.6705 -17.2774  17.8769
#>   -2.1787   3.2400   8.1352  -9.9455 -12.1601  -0.2829  12.6815  13.2585
#>    1.6151  -1.3260 -11.9900  -7.2061 -17.2329  -7.2095  19.5002   1.7998
#>   -4.0386   1.3766  14.6666 -14.9744   5.1421   5.7114  -5.1985  15.1602
#>   -0.2617   7.1993  -0.9817  -7.9246   1.7207  14.5220  -0.0675 -12.1556
#>  -11.1958  -6.3385  15.9153   0.6075 -17.9348  -8.3727 -10.2120   3.6211
#>   -1.0680   7.6759   3.6179  10.2442  17.4198 -13.1661 -18.9085  -0.9270
#>    5.1278  11.1736  -1.2423   3.2927 -14.1860 -14.9952  -1.3274  10.8326
#>   -3.0554  -2.0570   4.8884  -6.9017  15.7886  -3.5158 -11.9785   7.7030
#>    4.6972  -9.0010 -11.6275  14.6585  -3.9476 -14.7950   2.5397  -0.4877
#>    0.2753  -2.6062   2.9556  -0.5906   4.4503   7.0242 -17.2946  -1.2811
#>   -3.3044  -2.5172   3.5429   7.1479  10.2590  -4.3571 -20.3712   1.3735
#>   -1.7201   4.3875  -8.8817  -5.4367   9.7815  -4.7742 -11.6706   2.7750
#>   -0.9331  -8.3578   3.3899   1.4346  -3.0809   4.9186  -5.6419   5.7462
#>  -11.9297   2.7698   5.4346   8.0609   5.4111  -0.2548   3.6500  10.1236
#>    1.8372  -2.0063   1.6288   6.2580 -14.4055  -4.1152  -6.9689  -9.2841
#>   -5.0719  -3.7939  -7.9658   3.5784  22.2396   3.8050   2.2650  -4.3948
#>   -2.7945   2.4061  -5.8731  -3.5944 -11.6109  -4.6588  -0.5616  -7.8763
#>    4.1990   8.9611   6.6751  -5.6257   5.7568  15.7237   2.9145 -14.3550
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]