Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.6179   6.5569  -1.7017  -2.2905  -8.2119   5.2714   1.2902 -16.7012
#>   -0.2820  -3.1253  -3.2709  -8.2807   4.0751   3.6053  -0.0342   6.9349
#>   -1.0387  -4.4605  -9.9420  11.4382 -14.8737   0.2402  10.1421  -6.2845
#>   -0.6976   2.2568   2.7123   7.6078 -14.1751   2.2164  -6.0608 -11.9811
#>    0.8822  -2.2313  -1.4552   6.1530  -1.8392  -4.1413  24.4274   3.9931
#>    0.8854   6.7887  -9.9752   9.0474   3.6083  -8.9590 -10.6756  23.6180
#>   -0.2366  -5.7301   5.8784 -10.1482  -0.0374   8.4531   0.6347  17.7092
#>   -0.2527  -3.7544   5.5412  -2.6847   6.8217   1.9700 -17.0238  -6.0923
#>   -0.5119  -2.7402   4.0853   3.7276 -20.4674  11.8407   8.5679  13.2635
#>   -3.6105   4.8623   4.5858   5.3633  -7.6474   1.1384   6.0810 -11.6014
#>    0.3445   1.6386  -1.9135  12.7484 -18.7748 -14.7540   9.5140   2.3993
#>   -1.0092   1.5179   2.5467  -8.1785  -2.2933  -8.0207   1.6248   5.8444
#>    0.8068  -2.4602  -0.5738   6.2696 -15.6382  -6.1976  -3.2296  -1.4149
#>   -0.4756   3.0075   0.7793  -1.7691   8.8402   0.4129 -14.3785  -6.7351
#>    4.6594  -4.8601   3.0532 -15.4708   2.1918  -7.8947  13.5293   2.1067
#>   -0.3226  -2.3752  -1.2675  12.6183   1.2049   2.9623  -2.4678   1.7659
#>    1.0420   0.6596   2.5492   2.9981 -13.9017   3.5937   9.8316 -15.9773
#>   -3.1836   4.5211  -1.7200   3.2213  -0.4887  -3.8135  -7.9609   4.4142
#>   -2.7481  -2.8186  -6.9064   8.3401   0.2130 -12.9889   4.8060   2.0976
#>    0.1950   0.4657  -5.3865  11.8674 -12.8031 -15.5486  13.0159 -16.4171
#>    0.1837   2.5129  -9.7296  -3.2305  10.9238  -8.0990 -13.1938   1.3641
#>    0.3109  -2.8359   3.6881  -6.0539   9.9480 -11.1680  11.6988   4.8962
#>   -1.2775  -3.5189  -2.4800  -5.2163 -12.8323  -2.8318   0.2574  29.3142
#>   -2.2162   5.3881 -10.6719  11.1559  -4.8157  -0.6197  22.9096   6.7406
#>    4.0965  -4.1973   1.4100   8.1985   7.3136 -16.4865 -17.4933 -21.1110
#>   -0.6764   0.6182  -2.2382   6.6058   0.5516  -5.5825   8.0075   6.1295
#>    0.7483  -7.7584   6.7514  -3.9776   0.2333   6.3566  -3.6234   4.9857
#>    2.5326  -2.3450   1.5780   0.8455  -2.0361   3.0976  -3.0777  -2.0229
#>    2.6288   4.7708  -1.3828 -10.1743   6.2502   2.9490  -5.5655  -3.0345
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]