Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.3853   5.9809  -6.6126   4.9155  -0.1426 -21.1145  -7.7203   1.5489
#>   -1.0476   0.6483 -10.7933   7.8796  -4.6406  -8.3001   6.4857  -5.9036
#>    1.9208   3.5081   4.2458 -14.4484  -7.9925 -10.5928  -0.7017   0.9416
#>    1.5383  -1.6862   3.7216   6.7954  -9.3561  -2.6100  -2.3544 -14.4412
#>    0.7915  -8.4863  -4.3397  -4.6211  -1.3658  -0.2912   2.5386  -0.7045
#>    4.4776  -3.6956  -0.6735  14.3199   6.2469 -10.4923  -7.1853 -18.2550
#>    1.3937  -9.4874  -5.1854  -3.2772 -18.9502   8.4655  -9.0171   2.8158
#>   -3.0514   3.4158   6.1027  -6.9168   4.4551  -9.4974  -1.9260   6.5099
#>   -2.4764  -0.4833  -0.4238 -14.1414   6.3336 -13.7413   4.1855  -3.9749
#>  -10.6234   8.1611  -0.6296  -2.3407  -0.1968   7.3579 -17.5663 -21.6035
#>   -1.9202  -1.9582  -2.5911  14.6079   2.9830  -4.9304   2.7868   5.3894
#>   -1.3909  -1.2548   4.0534  -5.7088  12.0610   3.0597 -12.9007   7.2410
#>   -3.2369  -2.7691  -1.5747  -9.0206   1.6411  -8.9579  -7.2600 -13.4431
#>   -0.0728  -3.6011  -6.9700   5.8883 -11.2652  12.2000   9.9121  -3.9538
#>    1.1218   0.4858  -7.9424   4.6651  -5.5864   7.6388   3.4262  -3.8371
#>    1.2348  -8.0823  -5.0841  -3.6927  -8.0569   1.3591   2.0985 -15.4110
#>   -6.0824   0.4257  -2.2054  -4.3810   3.3249   6.0636  -8.9896   0.7687
#>   -6.4874  -1.3920  -7.7584  11.3035 -15.2901   1.6291  13.6473  -4.0221
#>    2.1816  -1.7024  12.1614  -5.5044  -4.9064  -2.9194 -11.5465  14.4499
#>    1.2083  -6.6592  -7.4238   4.4205   9.5580   2.3801  18.9483   2.6930
#>    1.8627  -5.6981  -5.3364  -4.9364  19.4521   2.5220  -1.4178   0.5397
#>    0.2905  -5.6495   3.7636 -19.0531   9.1262  -2.4027  10.6099  13.4417
#>   -4.0809  -4.3525 -11.6608  -2.8238  -0.6085  -2.4248  17.8994   6.3193
#>   -0.0934   2.9035   1.2113   7.4646   4.8211  -8.9357  -4.4116 -14.6311
#>   -1.9476   4.9713  10.3374  -2.4474   4.3953   4.0852 -13.5300  -6.4481
#>   -1.2326   2.6038  -2.3627  -1.1068   4.6980   1.0818  19.2719  11.6731
#>    1.4358   9.8756  -1.2186  23.9016   1.8077 -11.1064  12.2099  -9.0390
#>    1.4178  -7.9099  -2.4789  -0.4417 -12.2526  10.2928 -30.5898   0.9056
#>   -2.3138   4.4757   0.4209  -8.0694 -14.0979  -5.1313  -4.2113  -1.8193
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]