Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -2.7914e-01 -8.1866e-01  7.0128e+00  3.6746e+00  9.1538e+00 -8.8826e+00
#>   1.9895e+00  2.8442e+00  2.0145e+01 -1.1767e+01  1.0319e+01  1.0331e+01
#>   2.4187e+00 -5.2425e+00 -4.8309e+00  3.4467e+00  3.3376e+00 -1.2724e+01
#>   7.4632e-01  7.1199e+00  2.9377e-01  7.9794e+00  4.2700e+00 -9.4865e+00
#>   2.7221e+00 -2.2052e+00  6.8833e+00 -6.5023e+00 -3.8127e+00  8.8628e+00
#>   4.2758e+00 -3.4495e+00 -3.9648e-01  1.7385e+00  5.5207e-01  3.2067e+00
#>  -3.9108e+00  1.5745e+00  9.3695e-02  6.6760e+00  5.1440e-01  2.6796e+00
#>   4.0899e-01 -1.5231e+00 -4.1832e+00  9.8397e-01 -1.5494e+01 -2.2610e+00
#>   2.8188e+00 -4.0582e+00 -1.0970e+01 -3.0671e+00 -4.6758e+00  5.8801e+00
#>  -3.1199e+00  8.7980e+00 -4.7895e+00 -3.9483e+00 -4.6159e+00  5.3939e+00
#>  -1.7230e+00 -6.2798e+00  5.3762e+00  7.9225e+00  7.2055e+00 -1.5884e+00
#>   4.2383e+00  7.8067e-01  9.0304e+00  1.1851e+01 -5.3146e+00 -2.3585e+00
#>   8.5881e+00 -3.4571e+00 -7.9484e+00 -3.1012e+00 -4.0803e+00 -4.6943e-01
#>  -1.0762e+01  5.6695e+00  1.1203e+01  5.9165e+00  2.0984e+00 -8.2084e+00
#>  -4.9147e+00  1.0669e+01  1.0736e+00  6.8077e-01 -1.2232e+01  1.2591e+01
#>   2.6531e+00 -4.0298e+00 -1.0399e+01 -2.8045e+00 -8.6672e+00 -5.5118e+00
#>  -3.3598e-02  6.1508e+00 -7.1409e+00  4.5861e+00  8.8665e+00 -1.5865e+01
#>  -3.1998e+00 -3.0367e+00 -1.3033e+01  9.2684e+00 -9.3019e+00  8.4248e+00
#>   3.6436e+00 -5.6741e+00 -2.6080e+00 -8.5732e+00  6.1900e+00  2.1856e+00
#>  -2.3912e+00  8.2135e+00  6.5124e+00 -9.8439e+00  4.4278e+00  1.5032e+01
#>   5.0797e+00 -4.0374e+00 -2.5748e-01  1.0843e+01 -7.6460e+00 -3.6952e+00
#>  -6.7668e+00  8.5911e+00 -5.7231e+00  1.2953e+00 -1.0642e+01  5.6973e+00
#>  -2.1790e+00 -3.8760e+00  5.3451e+00  9.2555e+00  1.0918e+01 -8.3572e-01
#>  -3.1462e+00  4.0900e+00 -8.3856e+00  1.4617e+01  5.8399e+00  1.0533e+01
#>  -2.1412e+00 -3.4219e+00 -9.7427e+00  1.4065e+01 -9.8419e+00 -1.3649e+01
#>  -2.0289e+00  4.6607e+00  2.5785e-01  7.2023e+00 -7.9641e+00  4.0306e+00
#>   3.0895e-01  7.8017e+00 -1.1032e+00  4.1764e+00 -9.0354e+00 -8.0103e+00
#>  -1.4504e+00 -8.0959e+00 -4.0900e+00  1.4133e+01 -4.6530e+00 -4.4598e+00
#>  -1.3788e+00 -1.1139e+00  1.2733e+01  7.2509e+00  3.9340e-02  1.9416e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]