Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.8141  -1.8720  -2.6022  -2.7916   9.8826  12.3892  -7.5016  -6.2221
#>   -3.6277   0.3061  -1.5077  -2.4659   0.5289 -19.2534  -4.8155  -1.5825
#>   -2.5692  -5.7844   2.0832  -8.2025   2.0393   1.2972  -9.9355 -12.1484
#>    0.2521   3.0399   6.1649   6.6736   4.4301 -11.2747 -13.2765  -1.2982
#>    2.8343   0.7882  -0.7366   0.0046 -10.0089   4.2134  -9.9751   3.2166
#>    1.3260  -0.7840   7.8500  -7.2064  -7.4478  -6.2298   8.1713   6.0059
#>   -4.1520  -9.5962  -0.1642   1.9652  -6.6311  -2.4553  12.0584   4.3601
#>   -3.5934   2.3734   3.6898   7.2896   1.0970   0.9388  -0.6665   1.0812
#>    2.5472  -0.1617  10.9659   0.6345  14.4437  -2.6767  -2.7178   1.6841
#>    2.9769  -2.3374   3.0703  -3.3891  10.8680  -3.1835  -3.6194  -3.2513
#>   -2.9947  -0.6568   4.0140   3.9364   4.8269  -1.5188 -11.5870  -5.1319
#>    1.6615   1.2901   6.5983   5.6015  -8.5863  12.5924  -7.4159  -7.8438
#>    0.3631   6.5887  -1.0090  -8.0493   0.7858  -3.4781   8.3656   2.6314
#>   -2.0649  -2.1491   2.4715   0.6225  -1.6603   6.4900   6.4032   6.7825
#>   -0.6549   4.1314   7.4321  -7.0024   3.3119   9.5355  -2.8700   7.9174
#>   -1.1828  -2.6127   5.5436   3.3742  -1.5802   8.7678 -17.4127  -8.4202
#>   -4.3796  -7.1861  -1.0515  -0.4780  -2.8552  -0.7306  -8.6832   3.9887
#>    1.5704  -7.8502  -1.9547 -14.2795   9.2882   3.6733   8.3381  -5.3997
#>   -0.4600   6.5359   2.5955   9.5416  -1.6235   4.4801  -1.7926  -0.8092
#>   -1.1650   2.1575   0.4311   1.5978   5.0545  -3.2396 -21.7695  17.5731
#>    1.1009  10.2491  -2.5397   4.6413   6.6129  -4.8905  -1.0414   6.4080
#>   -0.0236  -0.9042  -1.6110 -11.1854  -6.2666   0.7074  -0.0261   7.4516
#>    3.3918  -2.3077 -11.5664  -3.7480   6.2836  -7.6784   6.1829  10.9309
#>    0.3716  -9.0986  -2.5327   5.4517  14.5273 -13.1595 -11.2740   0.4829
#>   -3.8090  -3.9160   5.1047   1.8612  -6.4373  10.6516  -6.5137  -1.3625
#>    2.4342  -5.8414  -9.5347   4.3449  -5.9888  -5.4868  -0.8132   2.0625
#>   -1.8682   0.5582  -1.8462  -6.6782  -0.4097  -5.2277   5.7862  -0.6086
#>    1.2657  -1.7835   6.5304   3.9905  -5.9196  13.0635   9.2158  -7.1086
#>    1.4105  -2.0760  -3.7667   5.4064   5.4084   2.2060  -2.7034 -10.4426
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]