Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -1.3737e+00  3.3388e+00 -3.5219e+00  5.5158e+00  2.9689e+00  1.2062e+01
#>   6.8367e-01 -1.3017e+00  2.6494e+00 -9.2599e+00  1.0037e+00  1.7573e+01
#>   9.1928e-01  4.9670e+00 -9.3580e+00  2.9890e+00  2.0790e+01  1.5641e+01
#>   1.0479e+00 -2.3854e+00  1.6364e+00  9.4572e-01 -7.6827e+00 -3.0392e+00
#>  -3.3565e+00 -3.7610e-01  2.8732e+00  1.6224e+01  5.2363e+00  1.7162e+01
#>  -3.2586e+00 -1.8142e-01  3.4246e+00 -9.0993e+00 -1.9113e+01 -1.0316e+01
#>   1.2346e+00  6.6031e+00 -6.6066e+00 -1.2714e+01 -1.0058e+01  1.2898e+00
#>  -1.4482e+00  9.8200e+00 -1.3228e+00 -9.9507e+00  1.1248e+01  3.6963e+00
#>   2.7964e+00  4.1183e-01  4.9315e+00  6.7375e-01 -1.3236e+01 -6.1601e+00
#>  -1.7183e+00 -6.6756e+00  1.1175e+01 -1.8608e+00  7.1989e+00  1.3225e+01
#>   7.1823e-01 -4.7438e+00 -1.4000e+00  1.0564e+00 -1.0809e+01  1.3198e+01
#>   2.5027e+00 -2.5315e+00  4.8192e+00 -5.8242e-01 -4.9500e-01 -3.5190e+00
#>  -4.2429e+00  5.4594e+00  6.5404e+00 -4.7940e+00 -1.1839e-01  9.6927e+00
#>  -1.7083e+00  4.5620e+00  9.8312e+00 -4.5493e+00 -2.0693e+01  6.0411e+00
#>  -4.5535e+00  5.8844e+00  2.0683e-02 -7.7368e-01 -7.3964e+00 -4.6712e-01
#>  -1.3238e+00  1.6311e+00 -1.8797e+00  7.9027e-01 -9.7360e+00  1.2007e+01
#>   2.6469e+00  2.5021e+00  9.2872e-01 -5.5018e-01 -1.3616e+01 -1.1008e+01
#>  -1.9467e+00 -4.7537e+00  7.6265e+00  1.2007e+00  1.1736e+01  1.5335e+00
#>   5.2999e+00 -8.8177e+00 -3.8281e+00 -3.7914e+00 -6.8201e+00 -5.5155e+00
#>   6.7133e+00 -1.7021e+00 -9.2446e-01  5.3381e+00  1.2593e+00 -2.6992e-01
#>  -3.7962e+00  1.5563e+00  7.9954e+00  8.0205e+00  7.3091e+00  1.3621e+01
#>  -9.0417e-01 -1.9699e+00 -1.0778e+01  1.3030e+01  1.5489e+01 -1.5339e+01
#>   2.9924e+00  1.8545e+00 -3.5167e+00 -7.1141e+00  1.5098e+01  1.0328e+00
#>  -4.2144e+00  1.2435e+00 -1.3124e+00 -8.1288e+00 -1.7630e+01  6.4608e+00
#>   2.0440e+00 -5.0450e+00  2.0468e+00 -8.9532e-01 -1.1094e+01 -8.2134e+00
#>   1.8349e+00  6.5889e+00  5.2230e+00 -3.5653e+00 -2.0095e+00  1.3788e+01
#>  -4.3466e+00  8.4348e+00  2.9447e+00 -1.2435e+00 -1.5928e+01  5.7243e+00
#>   8.5845e-02  4.0467e-02  6.4963e-01  2.2590e+00 -2.1887e+01  2.2574e+01
#>   3.9623e+00 -9.4012e-02  9.6795e-01 -3.2980e+00 -1.5256e+00 -2.0525e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]