Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.7789  -4.1941   6.3138   4.7925  -3.3477   3.1718   3.5288  -0.2967
#>   11.2201  -7.6794   3.7922  -3.8444 -22.3963 -12.4446  -2.6337  -6.0351
#>   -5.5192  -0.3025  11.3284   2.7624   9.6416  28.6218  -6.4194   4.8110
#>    6.2739  -2.9709  -4.2548   3.6579  -4.6344  13.2113 -16.5711   4.7095
#>   -7.5048  -9.0288  10.7390  10.2901  16.7696  18.0395  -2.9666   7.3951
#>    0.1875  -1.3049  10.0312 -19.4202  -6.8391  20.8952  -6.4712  -4.3980
#>   -3.2447   6.1987  -8.6450   0.8058  10.2949  -6.9056  14.9343  10.7179
#>   -6.5083   4.0955  -2.8202  11.2155   6.2622  -0.2166  -1.3747   1.2777
#>   -3.7570  -1.5616  -1.4159  -2.1540  12.7830   5.7234  17.0871  -0.4372
#>   -9.8505   5.7781   4.4764   8.3631  -0.5948  -5.9022 -13.7172  -6.7047
#>   -5.0552   0.6036 -19.0817 -10.0836   1.0712  -0.9653  11.8780  -6.9943
#>   -5.0478   7.7108  -0.7628   3.6400   8.4859  -0.1084   2.7105  -8.0714
#>   -5.6782  11.1424  18.3096  -5.2911  -6.9808  -5.5506  -5.4690 -10.9227
#>   -5.5869   7.4298  -4.8651   3.9757  10.3116   3.3758  -3.2666   8.8890
#>   -2.2710  10.6241 -18.4737   4.6483  39.3914  13.8526  -8.9551  -7.8191
#>    3.2712   4.1045  -8.8532   4.9674   4.3441   6.8666 -11.1663   3.1033
#>   -3.6161  -9.9034   6.6033   6.6051   0.6469  19.1214  14.9124   4.8108
#>    1.6776  11.6288  -3.5217   5.4109   9.5111   4.1600   7.9680 -19.9157
#>   -5.6152  -2.9657  11.2679   0.4194  -1.4395  -7.0838   6.1600  -3.0198
#>   -1.6530 -10.6386  -0.8798  -7.9451   2.3128   0.0697 -25.9237   5.7163
#>   -2.3885   0.0195   5.6398  -4.5829 -13.5051  -3.6324   1.1617  14.5559
#>    0.4905 -10.2395   4.1012   6.7141  -2.6216  -2.2135  -3.0567   3.5302
#>   -2.6076   0.4380   2.6859  -0.1491   4.1207   5.6035   3.3924 -16.2739
#>   -2.3987  -0.7193   2.0956  -6.7046   9.6548 -18.8736   1.5406   0.1825
#>   -0.6682  -2.0915   9.6968  -1.5968  -6.6763  -0.3564  15.6004   4.7945
#>   -6.5234   4.2355   7.8390   0.7104   7.0788  -4.3309  19.6918  -6.8344
#>   -4.2297  -4.2338   4.3729  -9.8831 -11.7935   4.4064   4.6350  -2.2533
#>    3.2156  -5.9640  -0.6626  -4.8580  10.2113  21.5859  -7.1541   8.0076
#>   -1.3346   3.1380   8.7846   5.3172  -5.4560   9.0568   5.3915   0.3403
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]