Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.1879   6.5978   8.5431   7.8896  -2.9137  -4.4394   5.4587   2.8356
#>   -1.1338   1.3069  -5.4661 -25.2827   4.8978  -0.7837   2.5260  11.4867
#>    2.4605   3.4691   0.0363   3.0181  10.2803  14.4844   3.6076  -0.8040
#>    0.5134  -5.6926  -1.6550   1.0169   7.1487   2.2168  -2.4254  -2.8964
#>    3.3122  -9.9014   0.6698  -8.5676  15.4010  11.6616  -0.4996 -11.5029
#>   -1.3807  -0.1485  -4.9284   4.9477   1.1815   3.2830  -4.0575   2.3989
#>    1.6970   4.1291  -3.4745   3.3368  -0.7915   5.7951  -6.7470   2.9447
#>    4.3118   3.3862   3.5918  -0.2989  -3.0775   9.5391   6.1529   0.6779
#>   -0.9664   1.1073   2.2405   3.7148  -7.3034  20.5632  -7.4278   0.6565
#>   -3.6874  -6.1006  -8.1654  11.3270  -8.4410   0.3397 -14.1405  -3.2215
#>    3.7862   1.3160  10.2889   5.4389  14.5203  -1.5375  -7.1139 -11.2373
#>   -5.5912   9.7695   2.6081 -10.0718  11.1150   0.3128  11.0818  14.0080
#>   -6.5792  10.3521  11.4219  -3.0829 -15.4335   4.5204   1.4197  12.6005
#>   -5.1649   5.2885  -9.9471   4.8565  17.4995   3.3152  -5.5958  -4.9531
#>    5.2858  -5.6176  -6.5099   0.9928  -7.6810  -4.1662   2.3717 -14.2289
#>    2.4114 -10.3665   9.2929  -5.8692   3.9279  -6.2498  -3.0580 -11.0596
#>   -3.7049   1.1562  -0.6073   7.2346  -1.0521  -4.5334 -11.3961   0.4764
#>    6.3298   3.3386  -1.7018 -14.7625  10.4829  -3.3855   5.5326   4.8944
#>    9.8986  -4.6723  -8.1087  -2.5593 -16.8259  -6.6774  11.2201  -9.0049
#>   -9.9664  14.2638   2.5422  -1.2653  -5.4909  -1.8966  -4.6641  10.2956
#>  -10.9921   7.8491   2.0976 -11.3267  12.0871 -16.3419  10.0124   3.9321
#>    2.7828   3.8180 -14.1227  -7.5455  -6.2884   7.0004  11.3563   8.9698
#>    4.8154  -9.3137  -4.4583  16.3752   2.0126  -2.2115   3.2624 -21.9543
#>   -9.4397  -6.9560  18.3427   4.3250  12.7999   6.2355 -17.4033   3.4753
#>   -3.0899  10.3682 -16.1536  -2.1877  15.3684   3.2742  10.1510  21.8128
#>   -0.3692   2.2471 -19.8873  16.6135 -11.2553  -2.7854   2.3249  -1.6363
#>   -1.2188  -6.1167  -7.7915   4.5853  16.8967  -6.6331   2.1681  -6.7828
#>    2.0649  -2.1351   0.4893   9.1837   6.4788  12.0034  -0.9900  -4.6791
#>    2.5149  -6.8364   4.7894   0.5502  17.8473 -11.2515  -0.7835  -9.7510
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]