Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.0448  -4.4740  -3.0547   4.4064   3.7594   4.4063   1.5270   9.2846
#>   -2.1685   7.9636  -0.1354   9.8127   4.1183  -8.5827  -8.7831   1.9483
#>    7.4149  -1.6477  -1.3968  11.6602  -0.3008   8.9514   8.8665   0.4691
#>    2.7967   2.5173 -10.3174 -14.7567   3.4535   5.6202  -5.5539   3.0893
#>    0.1849  -6.1899  -4.4867  -2.5549   5.8609   8.8555 -15.3859   6.7236
#>   -2.2783  -7.4181  13.9831  -2.1585   5.9243   7.1959  12.4551   6.3629
#>   -1.5612  -1.9753  -7.0651   4.6271  13.3917  -4.0548  28.2764  13.2744
#>   -7.4021   3.6676  -6.6671 -17.3287  -5.5602  -5.9746 -18.9660   9.6307
#>   -1.7013   5.0411  14.0595  -7.2776  14.9115 -20.0471  -7.1112   9.4078
#>   -2.1782   5.1439  -9.2229   2.1961 -10.8463   8.2556  -2.0333   4.1289
#>   -5.2954   3.5889   6.5465   4.2264 -12.3710  -6.9513   9.7621   7.3783
#>    5.0006  -0.3474  14.4983  -6.3132 -12.1103  14.8459 -22.3267   0.3628
#>    3.6018  -1.5803  -4.5740 -14.5307  -4.7138   4.3391 -19.9689  21.8197
#>   -2.1931   0.3237   1.7426  -6.5801   0.1948  10.0803  14.6230  -3.7877
#>    0.7731   3.1233 -10.5320  -1.5769  17.9183  -9.8325  -1.3844 -19.6657
#>    4.8231  -1.8244   1.2186  -1.3239  -7.9192   5.2438  13.7237   7.9290
#>   -6.7214   0.8489   0.3995   1.1863   9.4498  12.6873 -11.4544   1.4233
#>    2.7838   4.5356   2.2672   1.5646  -2.3302   3.7796   1.7525 -12.3350
#>   -2.9292  -2.4194   7.9950  -8.7250  -5.5636   0.3672   7.2892  11.2869
#>    4.3206   7.1869   4.2603  -5.0336   4.0552   3.7863  -4.8050  -2.6301
#>    1.0247  -0.7066 -10.9438  -3.1475   2.8838  -7.1747  13.6205  -7.1021
#>    4.2166  -3.9441  -0.0407  -6.0272  -8.1774   3.6013 -17.7242  -2.8395
#>    1.1803   3.7522  -5.7321  -7.9209  -1.1306  -7.9591 -10.2024  13.3536
#>   -1.2772  -2.0475   2.8827  -3.3318 -16.8782  -2.2287   8.3422   4.8663
#>    6.0074   1.6652 -12.6583   4.8940   3.3186  -7.9812   5.3464   0.8064
#>   -2.4680   0.4886  11.9707   9.7945  -0.4726 -10.5030   0.6944  -6.3623
#>   -3.8714   4.3637   3.1543   7.3021 -11.5327 -14.6207   3.7293  -0.6942
#>    1.0824  -7.9226   0.5952  -1.6240  -0.3796  -0.8234  -4.3718   1.7426
#>    0.9807  -0.1728  -4.8560   1.4977   3.7526  -9.7554   3.2793  -2.4713
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]