Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.9414   0.5202   4.2504  -1.0097  23.1265  -3.4512   1.7865  -3.8145
#>    0.9142 -10.7192  -3.9338   2.3315  -6.3193  15.3291   7.8434  13.8837
#>   -1.9315  -2.2022   1.4515  -7.4705  -6.4530  12.0900  -4.2286 -12.3974
#>    0.3126   1.1094  -4.9965  -6.7342   1.7758   5.2456  -9.5648   4.6457
#>    3.5549  -6.2672   8.7246 -13.7122  33.2658  -1.4010  -3.6654  -3.4335
#>   -2.3158   1.5610  -4.5662  -4.7643  -2.8655  -7.9243  -6.3126 -11.4021
#>    1.8620  -7.1881  -1.0739  -7.4665   5.8524   5.9987   6.5853   2.6536
#>   -1.1816   6.3618   3.1061   0.6189   8.9993   3.3413   4.2777  -7.1017
#>   -8.6828   0.9423  -1.4500   2.0933 -15.6059  17.1301  -5.1317   8.0477
#>    2.0809  -9.5397   7.5432   0.2394   0.5704  -9.6067  20.4994  -2.6340
#>    1.4793  -1.7409  -4.7221  -5.3959 -13.2539   9.4616   1.8308   8.0710
#>    1.2255  10.9707  -9.9600  10.2914   3.9406   9.8702 -20.3823   3.2374
#>    1.2366  -7.6735  -5.1680  -7.7501  12.1385 -28.6402 -11.8331   5.1516
#>    1.7681  -2.2906   2.0853  -5.4046   2.8652  19.3642  -7.4441   6.5955
#>    5.5019  -7.0204   4.8977  -1.5886  -1.7915   6.5298 -14.2921  -4.3603
#>    1.4745  -3.0984   0.3212   6.7076   2.0536   4.1652  11.5041  -0.5729
#>    2.4995  -0.2387  -8.7305  -7.5367  -4.9384   5.3659   4.3928   4.2457
#>   -1.6086   3.3203  -8.2613  -2.4256   1.6930   3.7180  -4.9484   7.7287
#>    4.0470   4.3513   0.1869  10.9001 -10.0428 -10.9392  -4.7573  -8.8111
#>   -1.3190   6.3069  -4.3253   2.1272   1.4564  -5.5522  -2.2559   0.9381
#>    1.7730   3.4770 -12.5176 -16.6781 -11.9915 -11.0998  -3.7623   2.8967
#>   -2.5685 -12.8847   3.8974  -1.8927   5.1323   4.8730  -4.1794  16.6322
#>   -5.7262  15.8168   2.5481   3.4955  -8.0990  -1.7227  -6.1464  -6.3933
#>    4.5838  -7.5015  -4.2797  -8.7568  19.9657 -13.4357  10.3041  -2.0579
#>    4.6050   3.0427 -11.0861   6.2449  -9.0909   8.3913   5.9439 -11.0033
#>   -0.9357  -2.6543  22.2573  -1.6941  -8.7950   0.4530 -17.2962 -16.0974
#>   -7.3499   4.0720   5.2812   1.8933 -14.6723  -2.2511  -5.9437 -14.5083
#>   -3.3371  -4.4167   8.3322  22.3820 -12.0771  16.2409 -20.7026   6.1649
#>    2.2271  -0.1426   3.4177  -4.0437   7.9622 -11.1239  -1.6825  14.6222
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]