Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.1837  -9.1246  -4.1281  -3.9260 -16.9474  -7.2714 -11.1534   6.1084
#>    5.1592   4.9051   0.5255   1.4153   7.7043  -1.0043  -6.0817  -3.3133
#>    1.5213  12.6236   4.0762  16.8752  19.0556   4.6360   0.9944  -0.7588
#>    4.2630   9.7622 -12.1331  -1.3714   8.4504 -10.9346  12.5029 -11.6042
#>    2.0510   2.2218  -1.0245  10.2384   4.3696  14.4500  -5.8368  -8.1327
#>   -8.2662  -6.5800   9.5767 -13.3680 -16.9936  -3.5691  14.0395   6.9788
#>    3.6387  12.9093 -15.9727  -0.5229   1.7329  10.5235   1.8696  -1.2209
#>   -4.0528   4.5547 -10.4768   4.1936   5.3421   9.3093  -9.6231  -9.9179
#>    2.2635  -7.6665   8.5032  10.4784   0.9677   3.7254   2.6696  -3.5600
#>    1.8016   0.0184   6.9439  15.6482  -7.1086  14.4414  12.6477 -10.6147
#>   -5.0304   8.3226   6.9618   6.5264  -6.2854  -6.5797  -3.8556  -7.4767
#>    7.8786  -2.6917   2.1915  13.8148  -9.7387 -15.6610   6.5359 -14.6745
#>  -13.0661  10.9655   1.7537   3.8978  -9.1313 -24.4205 -24.5437  22.2806
#>    7.5206   7.4379  -0.6842  -6.3825  -0.0814   4.2590  12.6286  -8.4419
#>    8.9273  -6.7482  -2.2829  22.3506  -8.8826   5.1300   2.0522   8.7775
#>    8.8949   0.6099  -8.2416  27.2975  -4.4442   5.0939 -19.1990   4.3315
#>   -1.4416  -1.7271  -5.5048   3.2887   7.9489  -2.9694  18.3602 -13.3344
#>   -3.3465   1.2090   7.0285   7.9395   8.6678 -11.3345  10.1323  10.9746
#>    5.9713   9.5906  -2.6835  12.4741   3.1000   3.0384  22.2696  -5.1768
#>    6.9837   1.7928   2.6748  18.8876  -5.6632   6.6578   1.1275  -4.3489
#>    8.5768  -2.1082   3.0799  -3.8997  -3.4727 -13.1516 -19.7337  10.1142
#>    0.3446  -5.9405   4.6459 -15.4753   4.4149   4.3438  10.2251  -0.0946
#>   -0.6759   3.3632   4.5321  -5.0708   8.3841  -4.5919  -3.7330  18.6714
#>    4.7506  -0.8801  -0.2964  -8.1563  -5.3901   7.8163  11.0381  11.0304
#>   -4.9203  -4.5905   6.7312   5.1690 -14.5326  13.1301  14.7062 -12.7573
#>   -1.8002  -6.5037  -0.1177  25.0386   1.6891 -12.6995  -4.8632  -5.8685
#>   -1.8026  -2.8182  -9.4311 -13.9619 -13.3901   1.3127 -11.8760 -10.3017
#>    0.8429  -5.2043   2.8558 -21.9003 -17.9599  -8.8877  -0.0258  -9.4716
#>   -8.3883   1.3518   5.1001 -11.1597  -5.0083   0.6199 -18.9090  -7.5513
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]