Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.2964  11.8861  -2.7294  -2.7447  -0.5884  18.1188   4.3321  -5.7357
#>    1.7004   5.5629   2.0619   6.7253  -1.9571   7.1494  16.5185   8.2808
#>   -0.0977  -0.2646  -6.4253  -2.1373  -6.2664   4.1083  -0.3170   2.3014
#>    6.3438  -4.6360  -5.5374  -1.3540  15.1155   1.1414  -6.1441  -9.6151
#>   -4.2166  -7.9382  -2.3963   1.2700  10.1262   3.8031 -17.5670 -11.8356
#>    0.4138  -6.3129  -0.3676  13.0974   5.3623  -6.4037   0.1199  -5.9431
#>   -1.2212  -1.8299  -7.1244  -7.4415  -2.4249   8.0520  -8.6831  -1.7834
#>    0.6856   1.3943  -1.6449  -0.3777   6.6990   8.5248  -0.9882 -10.5353
#>    3.2412  -0.3175   9.9208  12.6106  20.6190  -1.3279 -11.6412  -0.3827
#>    0.4092  10.9463  -1.7780  10.2894 -11.2182  -6.6481  12.0747   1.3330
#>   -4.6506   1.0135   8.8193  -0.6658   1.5444 -11.4506  -7.9366   6.8487
#>   -0.8169  -8.0448   8.8618 -11.0948 -11.6541 -10.0521  -6.4092 -23.5793
#>   -0.0717  -3.5056  -2.0831 -11.4851  -1.5902  -8.0247 -10.0579  17.9075
#>   -5.2658  10.6866   1.9201 -10.5605 -12.3828  -1.7221  15.7501   1.8145
#>    0.5007   6.7437  -2.9717   4.1586  -1.1743  -0.6682   8.1503   3.9802
#>   -4.6833   1.6308   0.2013  -7.1451  11.5557   4.6212   8.5317  -3.1110
#>   -2.6797  10.5189  -4.2531 -18.9302  -7.7679   7.6323 -13.0029   0.5290
#>    1.7721   3.1194  -5.4065   0.3016   3.7564   0.8150  -5.0065  -9.7597
#>    1.7188   0.0850 -11.6263   3.4607   8.3368   4.7095 -10.7290   9.6683
#>    1.5954   5.0356   2.6848   1.4499  -3.9216   5.5274  -1.6582  -5.3664
#>    0.7180  -2.4054   2.6928  14.7281  -8.9036   0.9036   0.8032 -21.5212
#>   -0.3768   8.0162   3.7931  12.2556  -8.9299  16.1536 -10.4045   0.6579
#>   -9.7510  -3.3573  -4.0652   4.6647   2.7808  -9.8467 -12.2996   0.5402
#>   -1.9017   2.4555   3.3116   1.6345  -8.3840   5.0497   1.1540   7.2951
#>   -9.5561  -0.9939  -0.4294  -4.0373  -5.3134  -0.6596  -5.8683   0.7793
#>   -5.4799  -3.3862   0.8052   9.7521  -9.8099  -2.2059   2.6044  -0.0132
#>    7.5918   5.9265   5.7415  -0.6696   0.8432  -3.5564   0.1879  -1.3632
#>   -8.0872   2.5428 -10.3942  -7.0896  16.0697  -3.2356  -8.6690   3.4854
#>    0.5396  -8.1617  -2.6969  -3.8114   6.7150  11.6000  -1.8375  -5.1446
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]