Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -6.2550e+00 -9.9623e+00 -4.0290e+00  2.1525e+01 -5.0140e+00  1.7778e+00
#>  -1.0363e+00  2.8339e-01  5.0025e-01 -4.7678e-01  7.5699e+00  4.5645e+00
#>  -1.4343e+00 -6.7961e-01  5.1788e+00 -5.3165e+00 -3.6452e+00 -1.9439e+01
#>  -4.0032e-01 -1.8598e+00  4.8760e+00  2.8240e-01  6.6962e+00 -3.8433e+00
#>  -3.2607e+00 -3.9339e+00 -1.4381e-01  4.4666e+00 -3.1717e+00  1.9713e+00
#>  -3.9378e+00 -1.0940e+00  4.7242e+00  5.8931e+00 -1.5005e+00  1.1996e+01
#>   4.9275e+00  3.2600e+00  1.0800e+01 -1.2076e+00 -7.0157e+00 -1.8949e-01
#>   4.5533e+00  3.1032e+00  8.7091e-01 -8.0237e+00  9.9329e+00 -1.0201e+01
#>   4.6739e+00  6.0317e+00  4.6822e+00 -1.3188e+01 -2.2648e+00 -5.7020e+00
#>  -2.6546e-01  2.6496e+00 -1.0100e+01  6.3116e+00  7.8433e+00  8.9710e+00
#>  -4.2438e+00  3.9822e+00  7.7304e+00  7.2295e+00 -1.2108e+01  2.7465e+00
#>  -2.0235e+00 -7.8832e-01  4.5910e+00  9.6734e+00  8.4919e+00 -2.5120e+00
#>   2.2258e+00 -5.3944e+00  4.7477e+00  1.0383e+01 -4.1485e-01 -1.8214e+00
#>  -2.0116e+00  3.6021e+00 -4.9015e+00  1.5987e+01  8.2259e+00  9.9385e+00
#>  -1.5972e+00  1.1327e+00 -7.3212e-01  4.4557e+00  7.0255e+00  1.4915e+01
#>   5.2131e+00  4.0758e+00 -1.9069e+00 -1.0021e+01 -9.0523e-01 -4.7592e+00
#>   1.7985e+00 -5.8845e+00 -3.6678e+00 -1.1188e+01 -2.8137e+00 -6.1760e+00
#>   4.8462e+00  3.6024e+00 -6.6658e+00  2.1346e+00 -1.7771e+01 -1.9491e+01
#>  -3.2190e-01 -5.9617e+00 -2.9319e+00 -6.9721e+00  1.9557e+00  9.1984e+00
#>  -1.9670e+00 -1.0817e+00  1.2817e+00  9.7677e+00 -7.5418e-04  9.4590e+00
#>  -2.3365e+00  2.5770e+00 -2.4985e+00  5.6582e+00 -1.1801e+01 -9.5629e+00
#>   7.2239e+00 -9.3800e-01 -4.7581e+00 -2.1529e-01  9.7497e+00  5.1603e+00
#>   1.0676e+00  2.3392e-01  6.7475e+00 -3.5831e+00 -2.7848e+00 -9.5676e+00
#>   1.7728e+00 -3.4782e+00  5.1761e+00 -7.7071e+00  3.1642e+00 -8.8341e-01
#>  -6.9686e+00 -1.0003e+00 -5.0857e+00  4.8023e+00 -2.3135e+00  1.0586e+01
#>   6.7252e+00  4.1116e+00  7.1578e+00 -7.5669e+00 -9.1341e+00 -7.2361e+00
#>  -1.7093e+00 -1.1956e+01 -3.4845e+00  3.2862e+00  1.5235e+00  1.0085e+01
#>  -2.9582e+00  4.7366e+00 -1.2644e+01  1.0743e+00 -3.8851e+00  1.8954e+01
#>  -9.2474e-01  3.7028e+00 -5.8595e+00  8.6367e+00 -8.8955e+00  3.3679e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]