Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.1100  -3.2536   5.7185 -14.0088   4.4078  -8.9692  -5.3090  -2.8742
#>    6.1481  -6.8454   1.8396   1.2118  14.5783  -4.9955 -20.3762 -10.0144
#>    0.5514   4.0030  -1.0308  -4.2416   0.8547  -3.6091   7.8445  -0.9745
#>    0.3362 -11.0197  -5.8410  -3.0808  -5.9707  -9.5443   1.9141 -16.6547
#>   -0.2364  -7.8798  -9.3515  -4.0460   6.0230   9.4725 -13.0716  11.7983
#>   -7.0176  -1.9593  14.0982   5.1346  -8.5628   7.0167   5.4572  -3.6454
#>    8.1647   5.3639  -8.7967  12.0496   7.0082 -17.1494   2.5831  -2.7879
#>    0.0350   7.6228  -4.7437   1.4587   4.3953   2.5254   5.3404   0.6573
#>   -4.3482  -3.6799   1.5588   6.8305 -10.9967   7.5386 -15.8082  10.8272
#>    3.2530 -10.5342  -0.5067  -3.8208   9.8300  -8.6422  -5.6303  -1.1087
#>    2.1753 -10.6718   2.8731  -3.3277   6.1987  -7.7157   2.1104  -2.0694
#>    2.9618  -0.2105   7.2129 -10.0892   5.6708 -11.4786  -7.1618  -3.3335
#>   -1.4986  -4.9396   8.4710   8.2036  -7.1266   6.1372  13.7678   5.6473
#>    7.1041  -0.7393   5.8465  -5.3736  -1.2082   0.6091   8.4300  15.1499
#>   -0.8352  -2.4428   0.2537  -0.9826 -14.0791   0.5451  -1.7329 -11.9717
#>   13.7422   1.1658   8.3040 -10.2795   1.2747   4.2519  -2.2029  -5.4370
#>    2.5129  -1.0342  -7.8936  16.6557   3.7876   0.9542   5.1063  18.4890
#>    0.8418   9.7946 -11.3537  -4.7126   4.6759  -3.1978  -3.7692   6.4056
#>   -0.0266   7.2677   9.5981   6.4268 -11.5326  -5.8130  -4.3199   7.9827
#>   -2.3861   1.5148  11.3552  -7.6430   0.5534  -7.2794 -11.1998  -2.5481
#>   -4.7808  -6.9854  -9.6718   1.4623   8.0937  -1.1284  -3.5616   5.4572
#>   -6.9105   2.1331   0.2886  -0.4215  11.8823 -11.5655   3.9969   4.8690
#>   -0.5704   6.4502  12.2036   6.6935 -14.5760   7.5706 -12.9642 -10.9590
#>   -1.6901   6.7341   2.0266   3.2469 -19.4429 -13.6805   2.2215  -2.8277
#>    4.0332  -2.0817   0.8645   1.3437  -4.8086   8.6357   0.1174   4.3406
#>   13.3258   1.2264   7.0494  -0.1096  -3.9547   6.7677   6.8124  11.4184
#>    6.3550   4.0176  -3.9183  -9.5355  -8.2498  19.2435   2.7893  14.0062
#>    3.4589   4.1237  10.0494 -15.9369   5.9160  -5.2124  -7.0236  -6.5439
#>   -3.8199  -1.1549  -6.9034  -5.5937  -8.0406  11.2389  -8.2405 -16.7569
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]