Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.0371  -7.1950  -3.9076   7.6822  -9.7129  -0.6252 -13.5513   1.8175
#>   -1.1538  -7.0648  16.0355   9.4066   1.0696   3.9918   2.6689   7.3914
#>   -0.3149   3.5183  -4.3969 -12.1381  -5.2365   3.2720  -9.2522   3.0012
#>    3.8870  -3.9114   7.5599  12.6755  -8.2881   5.5405  13.7791  -0.5903
#>   -6.7376   4.4600   4.2702  -5.6958  15.4515   2.9479   0.4989   2.1827
#>   -1.5770   0.0612  14.5549 -14.6421  -3.6787  -4.3551  -1.7479  -0.4260
#>    1.3396 -10.0941  -0.7909  -9.8740   2.6546   2.5919   6.0950  -1.2881
#>   -3.2851  -0.4359  -6.5068   0.6851  -9.1445  -1.2409  -3.1029  -0.8243
#>   -1.3229   5.8704   1.5015  10.5405   0.2847   8.1799   4.9455  -5.0864
#>    0.9909  -5.3709  -5.5860  13.6362  11.3524   2.6930   6.1088  -0.0897
#>   -5.3100   3.5003  -1.3707  -2.3557   1.3419  -5.9749  -2.1346  -1.4542
#>   -3.6488  -4.4155 -12.4539  -4.9505   5.3078 -20.1800   4.3764  -3.4147
#>    3.6242   2.3453   7.4158   2.1551  -3.8757  -2.3085   2.0266 -15.3281
#>   -0.1535  -1.0426  -3.3556  -1.8202  15.0079  -8.2394  -8.6348 -13.3478
#>    1.3468   3.3926  -7.0570   4.1534 -12.6415   2.4109   0.4970  11.2261
#>    5.9458   6.2184   2.1737  -4.2551  16.4270   4.1276   8.7968  -0.0663
#>    7.6397   4.3998   1.1591   7.1393   0.0552  -2.4370   2.9784  -4.5614
#>   -1.7372   0.1509  -2.6480  -2.7668  -0.1821   7.9657  -3.4120  -3.1913
#>   -0.1272   7.2538   3.8789 -17.7754   1.3801  27.6583  -1.9852   3.6214
#>   -5.7200 -12.5108  -3.5806  -2.6059  -1.2299  -3.9275  -6.8807  10.8841
#>   -3.6312 -12.4986  -4.3909   6.6985  -4.1919   5.2984  12.3083  -1.5128
#>   -1.7813  -3.5641  -3.9247  -7.2183  -1.7657  -2.1283  -5.3546  -3.6128
#>    8.5525   4.2850 -12.3476  13.3001   8.6156   0.8295  -2.6849  -6.2504
#>    0.4479   1.7554   7.6376   3.5374   3.4549  -0.6440   7.2564  -0.4858
#>    8.9490 -12.7006  -9.5833 -11.0795 -10.1382   5.7102  11.6451 -17.3020
#>   -1.2357   2.9562   4.8345  17.4178   2.7008   6.5665  -4.2805  -3.2843
#>   -6.6987  -4.0276 -16.7095  -9.7767   4.4917  -7.1056  -0.8978  14.2974
#>   -0.4976   0.6466  -2.2454  10.7660  12.2744   3.1756  -2.2550  -0.2756
#>   -2.2236   7.6371  10.3721  14.6997  -7.6229   1.5307  -5.5057   9.4585
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]