Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.2600   3.8319   2.2497  -6.4298   2.2509  -7.3155  -1.5272  -3.1029
#>   -8.0994  -0.5016  11.0055   5.2949  -0.2404   2.0149  -4.4892  -1.3160
#>   -1.8255   1.0151   4.2235  -2.0686  -1.0247   1.5277  -6.9173   9.7023
#>    2.7056  -2.6090 -10.2376   1.5592  -0.8725  -6.4048 -11.8400   1.3001
#>    4.6956   8.5227  -8.6902  -7.5395  -5.1443   6.2750  -0.5981  -9.3029
#>   -5.7371  -2.0080  -2.8498   4.2955  -2.6449   1.7770   4.4277  14.6920
#>   -5.2413  -6.1114   9.6882   3.1060   1.8258   2.7737   5.4625  -4.2817
#>    0.6541  -2.1610   7.7607  -0.5899   1.2635  -5.9205   7.9460  -1.1219
#>    1.4766   2.0358   4.1338   4.6282 -11.8962 -11.1129  -4.9629  -1.3630
#>    2.9424   6.4627   4.7736  11.8406   1.9414   3.7589  -1.5581 -12.8696
#>   -7.7183  -0.6164   1.6834  -7.2197   1.9533  -0.0494  11.5201   0.4601
#>   -2.6469  -4.7077   2.5161 -15.5817 -13.0983   2.2936   1.4762  -5.0497
#>   -1.9371  -8.0700   2.1149   6.8986 -14.8856  -6.6082   2.3180   0.0057
#>   -1.5705  -0.3336   0.2143   6.8129 -15.1344  -1.1850   2.1526   6.6498
#>   -0.2827   7.7881   1.0392   3.3060   5.7607  13.5780  -6.2873   8.4764
#>    7.2464   1.0123  -6.9646  -1.9072   3.3235   9.7836  -5.3179 -10.0594
#>    8.3094   2.1703  -1.2118  -5.0555 -14.5438  10.8576  -1.0836 -13.4199
#>    4.6636  -5.9550  -8.7298   1.8447   2.1049   8.2775   7.8806  -1.8438
#>    0.2044   9.2094   3.3527 -11.1497   1.2670  -3.0727  -6.5379  -6.5132
#>   -1.0000   3.4096   0.9197   3.0697 -10.2935  -6.6861  -2.4262  -3.0364
#>    3.4758 -11.4472  -3.9710 -10.9141   4.2261 -12.1226  -1.4050   2.8459
#>    9.0849  -2.6810   7.4740  -6.1077 -17.2163  -7.3925  -6.6754 -14.5215
#>   -1.5068  -7.7084   3.7652  15.0488  -1.9946 -14.4983   3.5320   5.2768
#>    0.7945  -6.3438   6.7147 -17.9480   0.3248   5.5578  -3.9668  -3.9338
#>   -0.0963   1.3921   1.6662 -10.2900  -7.8369  11.0073  -3.0465 -14.0364
#>    5.7992   2.1048  -1.9769  -0.8574 -17.9363  -9.8884  11.6564  -7.4763
#>   -1.8408  -6.5745   4.9771  -6.1593   4.2931 -11.8724  -6.9172  -2.1605
#>    1.6869   7.1835  11.0716  -8.1338   4.6218   1.3790  -8.2259  13.5691
#>   -3.4547  -3.3043  14.7004   2.3239  -0.6374  10.0752 -14.7851   0.4698
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]