Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.2712  -5.8574   1.0193   4.8061   7.4663  12.4302  -0.4499  11.0645
#>    5.9352  -4.7003   1.9895  -8.4019 -13.9001  -3.9332   5.0920   7.1892
#>    2.0017  -1.0520 -12.4450   4.4454   4.8303  -2.8703   7.1059   4.7563
#>    0.4192   3.4999  -3.2317   3.8282   7.8275  -1.3991  -5.7380   8.6345
#>    0.8463   3.2408   5.4875  -9.1947   5.3058  -4.2278  -0.4084   5.0892
#>    2.0165   1.1405   2.0403  11.6505  -4.3496   6.4620  15.1489 -13.2814
#>    0.1065   3.7530  -7.8471   5.5177  -9.3343 -10.5744  11.4387 -15.2701
#>    0.6750  -0.7573  -0.8739  -0.3558   5.1681   9.4399   0.2876   3.8667
#>    0.4755  -4.0345 -10.7004  13.2112   8.1405  -2.5114  15.1563  -9.3910
#>   -1.9247   1.5445   4.0703  -0.9882  19.9505  -6.6990   0.1213  15.3478
#>    3.6622   5.8327  -2.5744  -1.1623 -13.8484  -5.9875   3.7151   9.5436
#>   -4.1663   4.4951   5.1636  -5.3665   4.1448  -7.8435  -4.8097   2.6688
#>    2.3594   2.2447  -4.7544  -3.4903 -12.8401  23.8412   3.4463   2.1805
#>   -2.1958  -2.0758 -11.3904   2.7563   7.6324   2.2108  -8.6942  -1.6015
#>    2.9325  -2.0371  -1.4230   0.4934   4.5110  -0.1676   2.0919   0.6472
#>   -2.1856  -2.7832  -0.5186  -5.2678   1.4195  -8.9278  -6.7200   0.7059
#>    0.7407  -2.7810   1.4406   7.1098  -1.3170   0.2298  22.3534   4.2062
#>   -1.9775   0.2145  -1.4619  -2.0971  12.9969  -6.0918  -6.7886  14.5359
#>    0.8343   1.8210  -9.3145   7.5142  -8.2091   7.2737  13.3939  -1.7138
#>   -1.4264  10.4247  -1.4148  -6.4032 -10.5909 -16.9270   2.3527   2.5341
#>   -1.4462   7.3762  -7.0862   6.7755  -6.7066   8.6476   4.7250   2.3668
#>    0.4263  -5.2305  -2.6275   8.3226 -12.8450   8.2100 -12.7368   1.0928
#>   -1.0251  -2.7648   6.3561  -9.1748  11.3205  -2.5425  -5.0680  10.1811
#>   -6.4641   0.3282   1.4907   6.5460  -1.1584   5.3875  -6.7024  -7.3873
#>   -7.3569  -8.2230   2.8855  -2.7713  17.6736  19.1762  10.1557  -4.6533
#>    2.6365  -0.5038   4.5439   6.7865   6.6756  -0.3349   5.8295  -9.9391
#>   -0.8414   1.6377   0.9516  -1.8037   4.8207  -6.3203  21.9905  -3.9782
#>   -2.7724  -7.2570  -2.2733  -9.9292  25.4713   0.3740  -6.4230  -1.7950
#>   -1.0066  11.0698  -1.4277  -5.1921   3.8799 -22.8760  10.1929   6.0378
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]