Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  4.9004e-01 -2.2118e+00  2.6792e+00 -1.6869e+01  6.1747e+00 -4.6311e+00
#>  -3.1264e+00 -2.9013e-01 -1.0741e+01  4.0800e+00 -1.6994e-01  1.1435e+00
#>   2.3463e+00  2.9955e+00  3.7125e+00  4.0087e+00 -3.7581e+00 -1.2633e+01
#>  -6.7611e-01 -1.0998e+00 -1.8618e+01  3.5787e-02 -4.8582e+00 -4.3801e-01
#>  -6.9202e+00 -4.4304e+00 -1.2165e+01  1.1449e+01  5.4212e+00 -5.2725e+00
#>  -1.9164e+00  4.4470e+00  3.4905e+00  5.1454e+00  3.7885e+00  1.0082e+01
#>   1.2765e+00 -1.4707e+00 -1.6516e+01  3.8530e+00 -5.2787e-01 -1.7547e+01
#>  -1.0006e+00  3.6810e+00  5.1242e+00 -3.8120e+00  5.2835e+00  8.9663e+00
#>   1.7991e+00  7.1669e+00  2.4675e+00  1.9160e+00 -5.3891e-01 -1.5328e+01
#>   5.8463e+00  3.1378e+00  5.3914e+00 -3.5787e+00  2.7376e-01 -1.1462e+01
#>   2.6513e+00 -6.9260e+00 -1.0320e+00 -4.6726e+00 -3.3233e+00 -9.2492e+00
#>   3.4938e+00  6.8698e+00  7.0319e+00  8.5746e+00  1.1683e+01  3.9480e+00
#>  -3.3379e+00 -1.2346e+00  3.0833e+00  5.9210e+00  1.5926e+00  4.1122e+00
#>  -3.0180e+00 -9.8177e-01 -6.4661e+00 -1.2407e+00  1.4545e+01  2.6959e+00
#>  -4.6566e+00  4.7959e+00 -5.6771e-01  3.3409e+00  2.3841e+00 -1.6036e+00
#>  -3.6607e-01  1.4829e-01  4.1541e+00  2.7636e+00 -6.4288e+00  9.3996e+00
#>  -4.5864e+00  5.1968e+00 -2.2114e+00 -1.1964e+00 -3.1422e+00  1.6673e+01
#>  -3.3163e-01 -9.6609e+00 -3.0763e+00 -1.9579e+00  3.5893e+00  1.0568e+00
#>   3.5367e-01  6.0018e+00 -3.1434e+00 -8.8070e+00 -2.4524e+00 -3.6561e+00
#>  -4.2602e-01 -1.4523e+00 -4.3162e-01 -1.8752e+01 -1.5004e+01  1.4246e+01
#>   4.1669e+00 -2.0169e+00 -8.7205e-01 -9.7213e+00 -1.2106e+01  2.7200e+00
#>  -3.3988e-01  7.4480e+00  2.6007e+00  1.1263e+01 -1.0055e+00 -1.8734e+00
#>   1.8731e-01  9.2967e+00 -3.4955e+00  1.0202e+01  5.7811e+00  9.3200e-01
#>  -4.4004e+00 -1.1515e+00 -9.0439e+00 -1.2277e+00 -4.3675e-01 -1.6003e+01
#>   4.4706e+00 -8.9032e+00  6.5649e+00  6.0552e+00  4.8683e+00 -8.6147e-01
#>   3.8440e+00  1.7175e+00 -1.1398e+00 -1.0590e+01  3.9837e+00  4.3091e+00
#>  -7.3516e-01 -1.0358e+00  1.4719e-01  5.0449e+00 -9.8735e+00  5.3764e+00
#>   2.2569e+00 -7.4327e+00 -3.8193e+00  1.0902e+01  3.3323e+00 -1.2359e+01
#>  -6.8187e-01 -9.1371e+00 -1.4422e+01  8.1919e+00  4.9191e+00 -8.4104e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]