Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.2325  -9.0451  -1.7716  -7.2851   7.0689   1.4530   4.8923  -2.6086
#>    2.9505  -4.8917  -5.4952  -0.2715   1.4306  -7.1836  -7.5268  -4.6226
#>   -5.4832   3.4155  -2.0039  -1.1188   3.8548  -2.0849   3.1057  -6.7876
#>   -6.0448   5.5671   7.6476  -2.8671  -5.5961  -2.3502  14.4068  -1.9719
#>   -1.5611  -1.7496   5.2772 -13.1528  -2.3476   2.1350   9.2997  -2.9305
#>    1.2349  -2.1788  -2.8098  10.6269   5.8085  -0.0731  -6.1919   2.8600
#>   -4.3372   4.7955  14.2785 -12.7773   5.5343   4.8435   3.6547  -9.4104
#>   -4.4657   9.7516 -11.3621   6.6029   5.7522  10.5112  -8.9839  -3.3317
#>   -4.5442  -0.0217 -11.0109   6.6778   6.8242  14.2679   4.3866  -4.5084
#>    1.0352   3.1557  -1.4914  -7.3272   0.2884  -8.9958  -7.3875   3.1321
#>   -0.1158  -9.2830  -1.2980   9.0329  -1.7760   3.0058 -10.1159  -1.3297
#>    5.2861   3.1270  -4.9118  -3.8550   1.8295  -5.7077  -6.2361  -4.8916
#>   -2.1305  -1.4558  -0.1586  -9.1631   5.6182   2.8310  -8.6000   0.3611
#>    3.3418   3.0572  -4.0177   0.8815 -14.1205   5.9217 -17.3437  15.3367
#>    2.8935  -6.9772  13.9339  -1.3570  -8.1092  -0.0588  -8.6167  -5.4491
#>   -0.5100   9.4274   7.9863  -5.5739   2.3201 -11.6950  -7.0550   2.2242
#>   -4.6716   7.8784  -4.9708  12.2948 -13.9429  -5.3840   6.6130   2.9416
#>   -0.0092   1.3185   8.5584  -5.8974 -13.3634   4.6480  -0.5633 -11.3001
#>    3.8887 -15.3399   3.6052  -2.9824  15.7304   7.4015  -1.5147  -1.7443
#>    0.3040  -6.7364   1.5421   7.9967  17.4052  -5.5638   4.8751  -9.3163
#>    1.7503 -11.2921  12.7081   2.7923  -0.0125 -12.2146  -0.9016  -8.2934
#>    1.3649  -9.2079  14.9464  -0.5240  -4.4718  -7.4879  -7.8656  -1.1809
#>    7.7824 -11.3006  11.4835   0.0714   1.9080  -6.3379   2.3361   0.9866
#>    0.6569   4.1091 -12.3106  -4.0246  -2.5086  -4.3204 -10.2506   0.2839
#>    0.7141  -4.1035   6.0444  -0.0333  18.2134  10.5645  11.9091   8.2824
#>   -8.5713   7.4389  -2.6046   2.3439  -3.4495  -8.3361   8.5900   0.9656
#>   -0.9187  -4.0509  -3.5899  -3.4631  -2.0261  -3.7843  -4.9500   0.0098
#>   -8.1415   1.0398   1.4451  -1.8682  -1.7394   8.8574  -2.7731  10.5941
#>    5.0535   6.1733 -10.1797   6.0106   4.4522 -12.0297  -6.7100 -13.8507
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]