Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -2.6603e+00 -1.1935e+00  7.5941e-01 -5.7214e+00  1.7637e+00  6.5549e+00
#>   7.5264e-01  2.6173e+00  7.1136e-01  8.7963e+00 -1.2036e+01 -8.5074e+00
#>   1.3880e+01 -2.9746e+00  1.7230e+01 -4.0814e+00 -3.3608e+00 -2.5426e+00
#>  -5.3349e+00  4.7282e-01 -4.2065e+00  2.1131e+00 -4.6770e+00  1.6627e+00
#>  -1.2043e+00 -4.5293e+00 -6.2910e-01 -3.0128e+00  2.2461e+00  9.0498e+00
#>   1.6203e+00 -6.4339e+00  5.4651e+00 -1.3519e+01  8.3929e+00 -3.8326e+00
#>   6.3248e+00 -4.8863e+00  6.1830e+00 -3.3266e+00  6.2500e+00  4.9992e+00
#>  -2.8288e-01  4.8652e+00  1.3320e+01  2.9984e-01  1.2923e+01 -4.2026e+00
#>   6.6413e-01  5.0824e+00  1.1182e+01  5.6948e+00  8.5785e+00 -7.4301e+00
#>  -4.0486e+00 -5.5126e+00  4.0029e-01  1.9600e+00 -6.0928e+00  4.0963e+00
#>   3.2504e-01  7.5124e+00  8.3841e+00 -9.7211e+00  2.6798e+00  2.0205e+01
#>   1.2724e+00  2.2827e+00  3.0705e+00 -8.8655e+00 -7.6431e+00 -2.7264e+00
#>   5.3367e-01  2.9690e+00  2.7268e+00  1.0986e+00 -7.6277e+00 -9.6160e+00
#>  -6.3997e+00  4.5763e+00  1.7548e+00 -5.3399e+00  6.5243e+00 -6.4705e+00
#>  -2.7489e+00 -6.0614e+00  7.9202e+00 -8.7077e+00  1.1841e+01 -9.6298e-01
#>   2.7570e+00  1.5231e+00  6.4317e-01 -6.2863e+00  9.1150e+00 -4.3493e+00
#>  -2.5215e+00 -9.2310e+00  1.7408e-01 -5.3545e+00 -9.4289e+00 -1.1802e+01
#>   8.3645e-01 -2.4055e+00 -4.5468e+00  9.0182e-01  1.1200e+01 -3.7581e+00
#>   7.4136e+00  6.9646e+00  1.1970e+00 -1.7000e+00  7.4589e+00 -1.4816e+00
#>  -7.1985e+00  9.1768e+00 -1.2210e+01  2.6241e+00 -2.2508e+00  6.2733e+00
#>   2.2337e+00  1.9976e+00 -4.8561e+00  8.2931e+00 -5.8627e+00 -8.5730e+00
#>   5.6941e+00  2.5522e+00  4.2548e-01 -1.7031e+01  1.8825e+01  1.6588e+01
#>  -5.6585e+00 -2.7773e+00  1.4931e+01 -1.5035e+01  4.5565e+00  1.0878e+01
#>   5.4241e+00  2.5227e+00 -4.1735e+00 -2.2140e+00 -1.5313e+01  9.6648e-01
#>   2.2587e+00 -3.5703e+00  4.2501e-01 -5.3443e+00  2.5056e+00  6.8394e-01
#>   7.1689e+00  5.1360e+00  7.7993e+00  6.4849e-01  7.7371e+00 -1.3990e+01
#>   1.6497e+00  7.0980e+00 -2.9471e+00 -1.0150e+01 -9.4947e+00 -1.8082e+01
#>  -3.8412e+00 -1.5039e-01 -8.0846e+00 -3.8356e+00 -1.0833e+01 -4.1982e+00
#>   2.7356e+00 -2.7809e+00  3.2044e-01  1.6868e+01  1.5824e+01  3.7357e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]