Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.9196   1.5655  -9.2327  -6.5908  14.3538  15.0733   9.3840  -6.7320
#>    1.9647  -1.3799  -2.6238  -4.0840   2.7411  -4.7717  -0.3994   5.9384
#>   -3.4264   8.7949 -19.3640  11.2820  -3.8451  16.9458  19.4769   6.9137
#>   -1.7846  -7.8092  -6.9513   7.9163  -8.5165  -4.5950  -8.2555  -7.9638
#>   -1.3263 -10.2708   4.8612  -4.7046   1.0300  -2.6963  -5.7283   1.3505
#>    1.9930   3.9236   1.9400  -5.6588  -8.4683   0.6905   0.0188   0.7126
#>   -2.3321   5.2012  -1.1651 -10.1667   3.5308  -9.7719  -8.7888   7.6550
#>   -2.7895   5.3685   1.5993  -4.5086  -6.6737  18.7401   4.0411  -0.7163
#>    3.9937   1.0289   5.4708 -13.7884  10.5674  -1.5079   6.1496  11.9331
#>    2.2051   5.3505  11.1293   7.5793 -16.6945  -1.0038  14.7113  -8.6380
#>    0.6814  -0.7078 -12.0269  13.4103   9.3180  -8.6300  -1.1164  -2.1629
#>    1.0913  13.6954 -17.4187  -3.7671  -3.8521   2.1426   5.7767   3.1749
#>    1.3596   2.4984   7.3530 -11.2640   6.0303  -0.1125   1.9539   4.4817
#>    4.4543  -0.0885  -8.9564  -7.4650  -4.1818  11.6433   9.9584 -11.9100
#>   -2.6903   1.5458   0.1536  -3.4509   4.8472  10.5928 -15.9835   3.7030
#>    3.4548   6.4674   1.1910   6.7764  -1.7646  -0.9041   5.0166   6.9203
#>   -1.7737  -1.0331  -1.8366 -13.6703  -2.5770  15.5444   7.0070   3.0571
#>    3.7285  -4.6865  -1.1601  -1.6895   0.2441   2.6885   7.7933  10.4847
#>   -1.2373  -0.7281  -5.6988  14.1751   6.2250   7.9834 -11.6960   2.7164
#>   -1.3406  -5.2739   5.4916  -6.5660 -11.5061  -3.1177  -6.3802   3.3506
#>    6.3349  -4.9644  11.4142  10.5833 -24.2193  -8.1003   3.7874   5.5577
#>    0.1224  -1.7973   2.4695  -6.9530  -1.0061 -13.5242  -7.3947  -7.3807
#>    2.5478  -3.5256   8.9809  12.2055   3.9048   3.3523   3.5004  -2.8173
#>    4.8033   1.9659 -10.2418  -3.7853  -2.9302  13.1894  -7.3345  11.2719
#>   -3.9233  -5.5097   5.7301 -10.2651   4.6894   6.6486   1.6792  -8.8338
#>   -1.4128  11.1888  -5.5073  -2.1493 -13.2257 -18.4212   7.8300  -3.8917
#>    0.7830  -0.9869   5.6867  -3.8347 -20.5789 -15.8172  -8.7797  -0.5532
#>   -5.0278   4.2855  -8.0314  17.5553   0.6062   5.4031  -2.3348   0.4351
#>    2.3055   1.8314   5.4912  -3.1982  -3.0097  -6.0303   4.2863 -16.1112
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]