Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8 -12.4615  15.6209  10.7018  -8.1005  20.2285  -4.6768  23.6946  -9.1062
#>    7.9304  -4.2913  -0.5807   6.0667  -8.3172   3.2514   6.2887   5.5642
#>    0.8955  -4.7185   0.4786   0.2786  -4.2000  17.0799  -5.1935   2.4829
#>   -3.2433   7.1875   0.2157 -16.3468   0.6705   6.7666  -5.3991  -6.4917
#>   -1.7532   5.7043   3.4687  -5.3034   6.5297  -5.1223  16.0832   0.0753
#>   -9.7786   9.9117   4.3798   4.3155   0.3461  15.1693  -0.0273  -5.7618
#>    2.8974  -1.6017  -2.7876  -1.7270 -12.7784  -7.1501   2.8309  -5.7519
#>   -0.0740   2.0277   2.2942 -25.5474  11.6635  -3.4930   0.8733   3.0620
#>   -0.4164  10.8095   8.5107  -5.7920 -11.8497   3.9859   3.0670   1.2472
#>   -2.9690  10.4632  -3.7998 -18.4488   5.0878 -19.4267 -16.6951 -16.2803
#>   -6.1575   6.8091  -0.0122  -2.4474  13.3193  -5.0831  -4.4707  -6.8493
#>    3.2634  -3.5554  -0.5155  -2.6291  -2.2611  -8.1912  -4.8778  10.2639
#>    1.3981   4.0184   1.8875  -8.1810  -3.9452   0.7531   1.8719   2.7027
#>   -1.0008   6.1284  -5.8967  -1.0866  13.8073  -9.5134   7.0282 -23.9275
#>    2.0389  -3.0146   6.2437  10.6739  12.9336  -9.0924   6.7242   4.3572
#>   -1.4926  -5.8778   9.5575   4.4741  -2.4268  22.8406  15.2115   7.5091
#>   -0.9723   6.4417   2.6108  17.6531   4.5988  10.5801   2.1001  -0.1406
#>   -0.6730  -3.2506  -4.3189  -0.5666  -8.7902   2.9718  -3.3199   1.2019
#>   -1.3615  -0.3968  -0.8310   8.1256  -7.2125 -10.2870 -10.8905  -0.5163
#>   -2.8349  -1.5782   0.2748  -0.9416  -8.1241  -0.8654   5.5994  -2.2917
#>    3.6071  -0.6200  -1.2082 -15.8686   2.2813  -1.2548  -9.6540  -6.9913
#>   -5.3298   5.4207   8.1231   0.3379 -13.9211   8.3930 -15.3788   5.1333
#>    0.6763  -2.3973  -4.7267  -1.3385  -7.5817  -5.0071  10.6929 -11.5856
#>   -0.1948   1.0335  -8.0182   0.3433  -9.2623  11.1553  -4.9753 -13.8490
#>    7.7156  -3.0188  -4.2596   5.4530  11.1462   3.3635 -22.3047   4.9006
#>    1.5216  -4.0235  -5.4055   3.7736  -0.2430   1.1883  -2.7102 -10.1614
#>    6.6270   7.6024  -5.8185   3.5945  -6.7676   3.8273   0.1640  -3.1699
#>   -3.0859 -10.9131   4.6994  -4.9199   0.6072  13.9376   4.7516  -8.0755
#>   -1.9396  -0.4699   8.6271  -1.4344   1.4748  19.3399  -3.2989  16.2351
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]