Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -7.2670  16.5952  -4.8629  -7.0943   1.9678   2.5227  -4.9625   7.4989
#>   -5.4635   3.8903   3.8865  11.3790  -8.0820  -2.0897   6.3618  10.4869
#>   -6.0080   1.3584   9.8916   7.0975  -9.2405 -10.0766  -3.5749  -0.7304
#>    4.1109   6.1659  -4.9896   5.2957  15.1398  13.4701 -22.5350  -0.5174
#>   -0.6043   3.8609   5.4560  -2.3853  -3.4819   6.9116   4.0158  -3.6008
#>   -2.4865   4.4283  -9.3352  -1.6131   8.2094  -0.5051  19.3569  -6.8370
#>   11.2708  -6.1951  -6.5425   5.1174 -13.4427   3.8588   7.8997  -7.8339
#>    3.3267  -4.3267   2.7670  -7.9731  16.5378   7.7833  10.5985  -7.4056
#>    5.1914   2.6879   7.0796   0.8823  -1.2029   1.8895  -8.2528   0.0684
#>    0.7891  -1.3865   0.0122  15.1959   0.4954   0.5866  -9.4027   5.7178
#>    5.9198   9.9154 -10.7647  -3.9575   4.6906  -2.4457   9.3253  -0.3188
#>    2.5399  -4.5340 -14.1687   4.6704 -11.6962  -2.0876  15.1993  -4.1051
#>   -2.1321   8.9384  -5.0377   1.7310   7.7498  11.5309   1.0765  -5.8099
#>    5.9983   3.8418  -4.8075  -6.9192  11.1685  -3.2350 -10.2290   6.0237
#>   -0.7617  -3.4704  -5.0763  -7.6116  -4.1729  -1.4313   7.1687   8.3437
#>   -1.4419  -3.6128  -4.2401   1.0840  -2.1811  15.5251   2.9187  -7.8812
#>    7.3881  -1.7212  -1.2980   0.6752 -12.2801 -15.6574   1.2363  -1.3189
#>    0.7312  -0.4092  -6.3456  -8.2778  -6.1428   4.9383  -1.1979  -6.0026
#>   -3.9609  -1.3275   3.1854  -7.8386  -0.0675   8.5128  14.7451  -2.2086
#>    0.2034   7.3333   7.7074   9.7590   8.5098  -4.9038  -3.7887 -11.2257
#>   -2.2958  -1.4940  -9.7120  -6.5994   0.6263   7.3246   5.6939  14.3308
#>   -0.8035   1.1422   1.0380  -8.2179   5.2594   3.7187   8.1251   2.0947
#>   -4.2300  -1.0756 -14.9917 -16.5108   8.6013   5.4735   4.2512  10.5334
#>   -3.5379   2.8051  15.3376  -8.0159  -5.6255   9.9038  10.0590   6.5496
#>   -2.8921 -11.2132   2.0632  -0.9587  14.7661  -0.0627  -6.6810 -11.2378
#>    1.6446   1.1465   0.9111  -3.4239  -3.2024 -13.7913  26.3089  10.0685
#>   -2.2194   2.8978  -8.8648   0.8099  -2.1548  -4.7295   6.3113   0.0421
#>   -8.9016   5.4786  -3.8211   9.5678   6.8621   1.7513  -8.8845   5.6666
#>   -0.7236  -6.9250 -17.1476   4.3253  -3.6720   0.1453   3.4021   4.1852
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]