Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.8431  -4.5277   1.7078  -8.4379   6.2364   0.8618   3.2674   2.6476
#>   -2.1646   8.7403  -6.5668   1.8795   6.6292   6.3058  -9.0047   3.6513
#>    2.5004   6.9158  -7.5644  -6.2373  -1.9097  -3.8546 -14.5544   0.5199
#>    4.7308   4.5258  -2.1296  -3.3059  -0.8133   9.3319   3.4232  -5.2021
#>    2.2594  -3.2161   4.3062  -0.9834   3.8872  -4.8239   9.3881   9.0559
#>    0.1231   4.8974  -4.7468   3.1541  -0.9548  -5.7299  10.6756   3.9636
#>   -0.1636  10.0765   1.4119   0.8662   5.1337   0.4102  -1.8089  -0.2002
#>   -3.1814  -2.2992  -6.7347  -6.2719 -14.5191  -3.4133  -7.9544  -1.3920
#>   -0.1062   6.5784  -2.1617   0.0889   2.1640   5.3668  -3.3427 -18.1204
#>    2.3063  -4.0711   8.7234  11.8064  -7.5485  -5.8433  -0.5842  -3.1063
#>   -5.3977   5.7346   5.5007   1.2534  -6.7971  -0.6415   7.5231  -4.0114
#>    6.4483   3.3532  -3.1441  -4.9618  -8.1007  -2.7780  -8.2614   9.8224
#>   -3.0651   3.9554  -5.0276   3.6504   0.1450  -5.5214   9.6561   7.1790
#>   -1.3103   1.5414   0.7067   4.7270  -1.1140   1.4997  -0.1085   4.9571
#>    0.2297   5.6266  -5.6733  -6.7500   4.3766  -3.9202 -13.4177  -0.4278
#>   -2.7401  -5.6056   0.3313   1.1509 -10.1543   1.3034  -0.0117   9.5527
#>   -0.2753   3.6910  -7.1691   2.3767  -1.5048  15.9234  -3.9535  -7.7734
#>    0.8665   2.1761  -2.8047  -6.5223  11.2971   6.6763  14.1500  -3.1491
#>   -5.0950  -4.2059   3.6726  -5.7404 -11.2094  14.1030   6.8051   2.7407
#>   -0.0034   2.2126  -2.2561  -6.8021   8.1696 -14.8577  -7.8420 -21.6631
#>    2.1622  -1.4624  -2.3173  -8.8891   7.7924  11.0636   6.2531   2.6304
#>   -4.7437   0.9153  12.4648  -4.8691   4.7400   5.5788 -10.8156 -15.0029
#>    1.3513  -0.3952  -0.5190   7.5155  -2.1896 -18.5845  10.1623  -1.1157
#>    2.3360   7.9041  -0.5299 -13.2324  10.4694  -5.3222  17.0443  -3.6760
#>    3.7355  -3.0358  -2.9141  -9.2875  11.6475  -6.3769  -4.1727   0.1649
#>   -3.2818  -2.0223  -1.4205  -9.6411  -7.1073   8.6161  -6.4617   0.8823
#>   -0.9171  -3.5685   7.4673  -6.2050 -12.7719   5.0728   1.3943  -2.3030
#>    2.7164  -2.6730   3.1717  -5.6099   4.6783  -7.1818   1.9281  -3.8184
#>    0.6532  -4.9788   2.7542  -1.3369   0.0926  -4.1141   5.5900   0.7466
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]