Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.4531 -12.5427   4.2129  -7.1181   0.9720   2.0450  -0.1502  10.5239
#>   -3.4541  -3.9336  -3.2165   5.7131  11.6638   6.6790 -18.4742  13.7155
#>   -0.1554  -7.6506  -3.1142 -13.6978 -14.4010   1.4028 -14.5448  -8.1318
#>   -0.3774  -0.0157  -3.9558   1.6437  -0.0608 -15.1888   1.7378  -1.1638
#>   -0.5378   1.2118  -9.1922  -9.4544  -6.9864  -7.2615  -1.3823  -4.7806
#>    8.4330   8.3982 -12.8005   4.3853 -10.2388  -1.3371  10.9107 -16.0205
#>    2.4287 -11.7724   6.7832  -4.2190  -1.4260   2.4962 -13.7062  -4.4965
#>   -4.7985   1.2750  -0.7655  -2.1561  -6.3673 -11.9846   6.8992  -8.4679
#>   -4.5100   5.8953  -4.1654   5.7789   0.5926  10.5848  -8.6431  16.3332
#>   -2.2294  -4.8900  -5.2663 -11.0330  -1.1756  -1.1082   1.3917   3.5820
#>   -3.8833  -1.4291  -5.3198  -1.5626   8.0524 -11.1010   4.2982  -7.4171
#>    3.8973   3.7780  10.3672   0.5809  -3.9790   5.0868  16.6088   2.2324
#>   -0.1985   4.0412  -4.2246   1.5362   1.0121  -4.8776  -1.3692  -5.3938
#>   -0.7251  -3.9876   6.9877   0.4836  19.4876   7.1964  -0.0438 -18.7095
#>    0.4881  -1.3677 -10.0421   6.0305  -4.1308   6.9753  -0.4860   1.0189
#>    1.2278  10.0448  -1.9185   0.1682   3.0124  -5.8227   6.4307  -0.6473
#>    3.0788   3.7551  -2.3179  -0.3971   3.0560  -2.5326  -1.4901 -10.6606
#>    5.2415   5.6415   7.4963  -8.7525   0.5164  -1.2155   7.3282  -3.6952
#>    4.6730  -0.7870   1.3613   7.2374 -12.6828  21.3673  -5.7332   9.5113
#>    5.5357   4.8157  -6.5210   5.2555  -1.4872  14.6503  -1.9681  -6.6399
#>   -0.6045   3.8079   5.6922  12.4415  13.9098  16.6948   3.5391   1.1183
#>    1.7775   0.5325   1.7267  -4.8417  14.2915   1.4020  -1.9060  -6.9005
#>    2.4248   2.7109  -0.4835   4.6064  -9.2175  11.0913 -10.3299  -2.2613
#>   -0.2594   0.5072   2.5842  -6.4928  -5.5376  -6.2914  10.5947  11.6838
#>   -8.0956  -1.1128 -12.9891 -16.4397  10.2201  -1.9896 -17.1267  -7.9609
#>   -3.1201  -6.7513   3.9882   0.6914  14.6549   5.6511  -8.4680  -6.3099
#>    5.9834   2.8235  -2.9428   2.9050  -2.8362  -3.2092  -3.8635  -2.9352
#>   -2.0611   8.6976 -10.9808  -6.6335   3.6190  -6.1608 -14.7574   7.2658
#>    2.0302   7.8888 -14.9498  12.4168  -8.2663  -2.9416   0.2328  10.2991
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]