Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.1426   2.7579  -8.3702 -16.7452 -12.9304   7.0647   5.3543  -4.3274
#>    6.7362  -3.3631  -8.6237  -3.4089  -5.9535 -10.8441  -0.8510  13.9754
#>    3.9869  -0.3017 -10.7550   8.1101  -7.7937   4.9925  11.8620  -4.4435
#>   -2.6450  -4.3068   2.3619  -2.1184  -9.0963  -0.2911   3.3058   0.4754
#>   -5.1597   2.4540   6.7668  -7.9147  10.7541  -0.7118   7.3313   0.7330
#>   -3.8081 -12.0379  -6.7497   5.0998  -4.5072  -4.9364  -6.7530   5.0762
#>   -7.2971   6.3285  13.1218  -3.1278  -4.8030   7.8912  -2.9701   8.2945
#>   -0.3169  -6.4622   2.3188 -13.9386  15.3791  -1.5193  12.0434   9.8749
#>   -0.4747   7.2086  14.4653  -8.2103 -10.1576   1.6905  -7.0560  -6.1440
#>   -3.4097   2.1129   3.1811 -14.1297  -9.3945 -18.0334  -9.1186 -11.7152
#>    3.7755  -0.4671  12.4166  22.6451  -9.6536  -4.1000  -1.0326  -2.2420
#>    6.9953   2.4839  -6.5286   1.7511 -17.4078   8.3100  16.9372  -1.1362
#>    2.4662   4.6785  -9.5617   0.1380   1.0249 -12.3160 -13.9514 -10.3798
#>    6.1983  -7.6904  -2.9331  -2.1549  -3.6702 -22.6316   6.0482   4.4482
#>    4.2767   4.7651  -1.5778  -0.2888   1.3816   8.5795 -14.2421  -6.2550
#>  -11.5738  -1.7334  14.4696  11.2630 -18.9867  -7.2256  -7.6779  -8.4609
#>   -1.6189  -2.7045  -0.5218   2.3316  10.5594 -17.1204   3.6473   8.5982
#>   -7.8243   6.4078   8.8893  -7.7022  -2.3499   1.4724   8.5730 -19.1269
#>    2.2712   0.3065  -0.1369  -4.1685  -1.6483  -8.3822   4.5285  -5.1143
#>   -1.8302 -10.9812  11.6061   3.7811   2.2985  -5.0946   1.3524  10.7755
#>   11.0660   3.4346   1.1460   5.4612   4.8348 -13.3078   3.6051  -2.1677
#>   -2.7555  -0.2232   5.3628   4.5534   4.8326  10.7736 -10.0445   6.5768
#>    2.3532  13.2538  -3.5607   5.0704  14.3635   5.1865 -17.2889  -4.8881
#>   -0.1623   0.7518   7.7387 -16.6984   5.4984  -5.8467   4.5688   3.1239
#>    5.1507  -4.9590  -3.5268   5.7569 -19.1993   2.3679   1.5085 -13.0656
#>    9.1532  -5.8386   0.5448  -3.6810  -3.7127   0.3264 -10.1684   1.0916
#>    1.7246   5.8925  -2.7110  -9.9635  -2.5008   2.0266  -4.1910   1.8566
#>   -4.6432   6.7031   2.7399  -3.2320   0.5903  16.2802   2.1986  11.9252
#>   -4.0702   0.4652  -2.4096  -8.6690  10.7675   9.1719  -8.7226  -2.3165
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]