Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -7.8922   9.2365   7.3576  15.5201  -9.8887   0.8994   3.6535  -2.6025
#>    1.5050   5.4529  -1.5455 -25.3313 -12.3474  -3.9943  -2.6968   1.7907
#>   -6.1760   8.0905  -2.9111   7.8834   2.3645   0.9987 -14.5170  15.1788
#>    2.2996  -1.8616   0.9384   7.1320  -2.6024   9.0788  -0.2553   7.2394
#>   -6.3939  -4.9864   1.8641  -7.3953  -4.8291  -8.0913  10.7904  -4.9091
#>    0.6440  -7.0944  -8.4458  -1.7289  14.5293   0.6341   8.0029   0.0324
#>    3.2303   6.3309 -10.0062  12.3253  -6.6611  15.8269 -13.2482   9.8936
#>   -1.3587   7.9388   0.1383   6.9223   0.8657   9.8325 -25.8169  17.1190
#>    4.8419  -8.9290   3.8659 -12.1332 -12.9278   8.6377  -0.1414   7.0772
#>    2.6511   5.3780  -2.3615   5.6949  -2.8841 -14.8556   5.5766   8.2805
#>    3.1888  -0.4384  -1.0740   1.2358   7.4534  -1.1734  -0.6463 -14.4253
#>   -6.3769  -9.3538   2.6740   7.3244   5.4845  -0.9019   1.6763 -13.3579
#>    5.3278 -12.1345  -8.0627 -15.7613   8.5869  -8.1443   5.0156  -9.5959
#>    4.4660 -11.0167   0.5863  -5.8302  21.6646  13.0303  22.3384  -0.8312
#>   -2.4028  -1.0278   4.6037 -14.1590  -1.6189 -11.5605   5.3412 -15.3427
#>   -6.2193   0.9846  -0.7301   3.7853 -16.7334   4.4007   3.9683  -1.8185
#>   -2.1866   5.2674 -14.8103 -11.2159 -16.8844  -2.2047 -12.4383 -16.7771
#>   -4.9932   0.2471   7.7365  18.1610  -4.2052   1.6081  -0.0407 -10.4134
#>    3.5200   0.9713  -0.4222  -8.1364  -4.8588  -4.8802   4.3470  -7.0135
#>    5.7772   1.5006  -2.6001   0.8241  16.7907  13.3687  10.3375  -2.7505
#>    2.3970  -4.3873  -9.5275  -3.4568   8.2159   9.3348  20.4831   1.5156
#>   -3.4627   2.6852   3.6834   0.8826   6.9499  -4.5705  -6.7169  20.1141
#>    4.8715  -5.4444  -1.3803 -15.7234  -2.5747   1.2057  -7.3791  16.8661
#>   -1.6734  -4.8367 -17.2774   1.7562  10.8501  -3.9920   9.7508   8.6371
#>   11.6076  -7.0359  -5.4778 -10.8758  -1.0508  -7.3860  21.5847  -0.3113
#>    0.4179  -5.7815   3.5347  -2.2070  13.5157  14.0270   4.5319  -1.9676
#>   -1.0603  -4.7533  -2.7666   1.5129  16.1075  -0.0457  -3.4521   9.0375
#>    0.7942 -14.2032  -0.3368  -8.3866  -8.3915  -6.3281  -2.7678   2.4018
#>   -1.1306   2.9425  -2.0303   8.3395  -5.5211  10.7491   8.2081  13.0103
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]