Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.7167  -3.7759  -6.5194   8.2110  -3.7002   1.2415  -6.0168  13.5475
#>    5.4922   8.4175  25.9147   3.6569  15.8802 -21.1239  -2.8959   1.1383
#>    3.8449  -3.9025   8.4641  -0.7739   8.6294  -3.9442   0.3415  13.2660
#>   -8.8816  -6.7987  -4.6795 -27.2260   6.1418  13.6678  -1.7330  10.9267
#>    2.8295   2.8454  -5.3050   2.0277   2.0823  -3.2495  -4.2813  -3.8769
#>    4.7718   8.0102 -17.6877   8.6513   3.5385 -18.3072   4.5045   2.2173
#>    8.9833  -4.8811  -0.9484 -23.5667  -9.5290  17.5106   8.1919   3.5581
#>   -5.8732   1.5384  -2.9948  17.1652  -6.1655 -26.7523  -4.4470 -28.6713
#>    3.2285   0.6553  -3.0953   4.9807  -1.4156  -8.1941 -10.6332   4.5426
#>   -3.7026   3.7639   1.4077   2.3540  -2.9339  -0.5669   6.8801   7.1702
#>   -9.4506  13.2466   8.5967  11.6993  -2.9693  10.0425   3.8625   2.5689
#>    5.7072  -5.6581   5.0532   1.8888 -29.9156  -5.9075  -1.1630  13.8934
#>   -1.0044   7.3542  10.8784  -6.8679 -12.8202   5.4008  -1.7685 -11.7020
#>   -0.0909  10.7359  -0.7677  -8.7003  12.6508  -3.5108   2.6094  -0.2405
#>    7.5106  -2.0164   3.0953   0.9762  18.6453  -7.2243  -2.9015 -14.8094
#>   -0.0740  -5.1327   2.7431  -5.0395   2.8286   2.8159  -2.3712 -15.0118
#>    2.2782  -2.4544   3.0339  -6.1869 -12.6460  21.0073  14.8598   2.1180
#>   -0.4457   8.3969  -2.3834   1.5012 -10.8078   2.6022   9.5525  20.6374
#>    8.7625  -3.2480  10.7423  -6.5226 -11.2905   8.5412   9.3938  12.2866
#>    2.3719  -2.5002 -22.6699  -5.1378  28.4061   7.7861   3.3336  -6.2052
#>   -2.4062   8.1566  -4.2176  -0.1826  17.1049  -3.1028  -6.2275  -3.4257
#>    0.3898  -4.9946  -2.2377   4.9842   5.1123  22.7730   1.6080  12.3250
#>   -3.8689   5.3910 -11.5330   1.3869  15.2529 -16.8506   9.9313 -13.6264
#>   -0.8010  -2.3734   3.1835   2.3756  13.6392   6.6837  -5.3145  10.8942
#>   -1.9767   1.3784 -10.4292  -0.0807  15.8015   5.7135  -3.7850 -20.6307
#>   -2.9300  15.3530 -12.9037   4.3453   7.0884  -8.7718  -6.7811 -14.4864
#>    4.1759  -6.4659   9.8169  -0.1350   4.9076  -8.5934 -21.9274   2.3579
#>   -4.0099  -2.5341   3.8247 -13.7346 -18.5851   9.6251  19.2981  -5.3652
#>   -5.0210   1.9389   2.8769   1.3929   5.4961   3.4080   2.8252  11.2336
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]