Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  1.3792e+00  1.2188e+00 -6.0411e+00  8.5078e+00  9.7778e-01 -1.0451e+01
#>  -1.2124e+00  3.9495e+00 -2.7124e+00  8.5176e+00  5.3384e+00 -2.0434e+00
#>  -1.0386e+00 -2.5671e+00 -4.0331e+00  3.8127e+00  1.0786e+01 -1.1057e+01
#>   2.2333e+00  4.9126e+00  1.1322e+01  3.5338e+00 -1.7060e+01  4.2108e+00
#>   2.2589e+00  4.0674e+00  1.5682e+01  6.0481e+00 -1.5981e+01  1.1253e+00
#>   1.5824e+00 -2.0512e-02  1.4946e+00  1.0008e+01 -1.4802e+00  1.5871e+00
#>   4.1652e+00  8.1240e-01  9.2885e-01 -2.4076e+00  3.8860e+00 -2.8095e+01
#>  -1.4552e+00 -1.4172e+00 -1.7308e+00  1.5249e+00 -7.6638e+00 -3.0284e+00
#>   3.7711e-01 -1.6862e+00  3.6096e+00 -6.7953e-01 -1.1425e+01 -2.5880e+00
#>  -3.2360e+00 -3.8266e+00  1.1265e+01  4.3026e+00  1.7749e+01  7.9974e+00
#>  -1.8190e-02 -3.6392e+00 -1.3128e+00  1.0980e+01  1.5351e+01 -1.2478e+01
#>   4.4917e-01 -6.5852e+00  9.1600e+00  7.2908e+00 -9.4806e+00  1.7313e+01
#>  -1.0566e+00  4.7586e+00  3.8963e-01 -1.9323e+01  8.9337e+00  5.2114e+00
#>  -2.6098e+00  1.6400e+00 -1.0974e+00 -3.0780e+00  1.1050e+01  5.9513e+00
#>  -4.8043e-02  6.2977e+00 -8.9546e-01  2.0021e+01  6.3458e+00 -1.1264e+01
#>  -1.5212e+00  2.1745e-01 -6.6272e-01 -1.0822e+01  1.2411e+01  2.0700e+00
#>  -2.6006e+00  2.1662e+00  6.6137e+00  3.3443e+00  7.0459e+00 -9.1164e-01
#>  -1.8745e-01  1.0485e+00  1.4279e+00 -1.0173e+01  3.0753e+00 -7.3786e+00
#>   1.8950e+00  4.0205e+00 -6.0826e+00  9.6373e+00 -4.1691e-02 -9.4924e+00
#>   1.9659e+00 -7.6649e-01  3.8977e+00 -3.8597e+00  9.6512e+00 -3.1348e-01
#>  -4.5931e+00  3.2540e+00 -1.5030e+00  3.1524e-01 -8.7863e+00 -8.0319e+00
#>   1.0175e-01  6.1535e+00 -3.3728e+00 -1.7143e+01 -7.9095e+00 -4.3702e+00
#>  -5.6170e-01  1.0969e+01  2.9395e+00  1.1344e+00 -8.0580e+00 -4.7211e+00
#>  -7.2072e-01  2.3061e+00 -1.0911e+01 -5.1249e+00 -9.6263e+00  7.9419e+00
#>  -2.5468e+00 -4.6781e+00  2.0741e+00  1.1801e+01 -2.9492e+00  1.2864e+01
#>  -1.0929e+00 -1.8819e+00  5.5026e+00  2.9863e+00  1.2355e+01  1.1777e+01
#>   4.8285e+00  5.5095e+00 -1.0227e+01  1.5284e+01 -4.7016e+00  4.5540e+00
#>  -2.6554e+00 -2.5120e+00  1.0433e+01 -8.0797e+00  1.0227e+01  4.8755e-01
#>  -2.9112e-01 -6.3570e+00 -4.1941e+00  1.5208e+01 -7.5456e+00 -9.1533e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]