Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.6009  -3.0907   6.7048  -5.3351   3.4108   7.5205  -2.1001  11.4293
#>    2.9074   7.4036   7.9261  -0.6546  -7.2050  16.3348   0.5399  -2.4014
#>    5.1057   3.8486  -2.6189   6.2165   1.0941  -6.2129  -5.1235 -11.0302
#>    0.5969  -5.5839 -10.4532 -14.4082  11.5065 -12.6085   8.6992 -12.9170
#>   -0.7321 -10.4049  -4.6968   7.1661   1.7653  -0.0295   5.4529  -0.2892
#>   -4.6075   6.4029   1.5999   6.4231   3.4228  -1.6054  -5.4020   7.3867
#>   -0.6051  10.2385   4.0239   3.4026  -5.5840  14.2716  10.2325  -5.5019
#>    1.0682   1.6124  -4.3045   0.4500   3.8420   6.4276  -6.8517  -6.8311
#>    3.4877  -0.8391  -8.3907   4.2054  -2.3401  -8.0738   7.4705  11.0660
#>   -1.8263  -3.4258   2.4065  -1.1189   3.1693   8.2028 -12.7542   3.5786
#>   -3.1347   0.9528  -0.5127  -6.6526   9.5016 -14.5412   6.5179  -1.1062
#>    0.1158   3.3181   2.5195   2.4969   1.3206 -13.2082   0.2387 -17.0395
#>   -2.8646   6.2529  -4.0109  -3.0115   9.6918  -9.4542   2.1988  -8.4941
#>   -0.3168 -13.2397   4.5012  -0.4156 -11.4620  -8.7095  14.1194   5.0855
#>   -0.4865   3.0043  -8.6639   9.4162   8.1065 -10.3952  -4.2042 -24.3807
#>   -1.1172  -6.6055  -1.4901  -2.2118  -4.9235  -1.9229  12.1827  -6.5128
#>   -2.0964  -5.5427   2.7605  -6.4746  -3.0387  11.1989  -1.2248  -8.3120
#>    0.2824   3.6702   9.1924  -2.4958  15.4752  -2.4117   1.6367  10.1871
#>    0.7950   3.3748  -5.6526   1.8643  -6.5943  -2.6598   5.0007   0.9657
#>    0.5117   2.3701   8.8831  -2.0205  -0.4719  -2.4485   1.6630   9.9582
#>   -0.8569   5.1537  -4.9864  -7.0192   3.9274 -17.1676  10.1120  -5.9565
#>   -0.4360  -0.4828  15.8336  -4.4724  -0.1477   3.6285   6.0099  -0.6469
#>   -0.7983   0.4974  -2.8639  -7.8047  -2.4996   9.2459   0.2395  12.7129
#>   -5.4476   4.4065   8.1019   6.9377 -11.4161  -6.5216  -0.8959   0.3961
#>   -0.7828   0.0729  -1.4849 -15.6637  -5.4276   0.3784  -6.0319  -2.1454
#>   -5.1700   1.8989   3.3202   4.6075  -6.5197   5.0535   2.1598  -5.2715
#>    3.0287   5.6549   2.9468  -1.4255  12.3457  -5.5039   3.6204  -4.1554
#>   -2.9843  -3.0265  -2.5511  -1.8575  -6.9759  -1.5870  -5.5869   0.6415
#>   -7.6079  -1.4024   6.0519  -4.2503  13.5191  13.5132   8.0399   3.1031
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]