Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.3660   5.5626  -2.0304  -2.6227  -9.9864 -14.3188  -0.8433   5.2373
#>    0.0493  -0.0058  -7.6052  12.1781 -13.4381   6.9510  -3.7032  -8.9581
#>   -0.3862   6.1954  -6.0315  -7.9175  10.2627   5.3010   0.4831  -5.9443
#>    3.2163   4.0663 -12.7927  -9.4972  -7.9097  -8.4615   7.3351 -11.8533
#>   -6.4583   5.6982  -2.3096   0.3100   9.7280  -1.6210   6.2367   4.6630
#>    2.7052  -0.0879   6.2182   0.1322  13.4369  -0.9066 -14.3779   9.5218
#>    6.1319  -1.0096 -15.7122   3.2740   5.7348  -1.6023   3.9480   5.2570
#>   -5.4651 -12.4261   3.5594   7.3103  14.2969  -7.4410  -8.3447   8.5984
#>   -5.7741   3.6122  -2.1066  15.8994   3.5974 -11.4112  19.0900 -13.1606
#>    4.7971   1.9323  10.0169  -9.2109  -5.1504 -10.0638   4.8716  -9.3736
#>    4.4440  -3.7536  -3.0887 -15.1787   7.1380   3.5181  13.3127 -13.1864
#>    2.8217  -7.2889  11.9525 -14.1273  -5.2670 -17.1256  11.9975  13.0676
#>    5.4604  -0.7160  -0.0910   4.0863 -12.9259   6.8274  -3.2938   1.8400
#>   -7.0470  -4.2094   2.8481   4.6641 -14.2529  -2.4329  -5.1320 -18.9428
#>    0.1580   4.2793  -9.7631  -4.4381   9.9994 -17.9878  -2.9291  -4.5583
#>   -3.6932   9.1590  -3.6979  -2.6076  -6.2823   3.2158  17.2903 -14.9355
#>    3.0845  -3.0992 -11.7184 -11.6004  15.9590  12.2631  -1.3854  -2.2587
#>    7.6046  -3.4009   2.5236  -6.7201  -1.5369  -9.3903   0.2653   6.6836
#>   -0.6416  -1.9704  -0.2304   4.9213 -14.9646  -6.0350   4.4375 -19.2357
#>    4.8606  -7.1361   8.0905 -18.6082 -13.9150   1.9641   7.6182  15.2569
#>   -6.9276  -4.2748  12.5706   1.0651  22.5987  10.5538  13.9224 -13.4727
#>    2.1932  -1.1622   0.5044  -2.9980   5.1772   1.8331  12.7444  11.3613
#>    5.7655  -3.8724  -6.1852   3.4930  12.7491  -7.5499  -6.5545  -7.1564
#>   -6.9049  -2.6433  -2.3763   3.1524  -7.2764   5.5545   2.2477   2.8561
#>   -1.1486   0.2625   6.0254  21.3822   2.1329   0.2466  -3.5659  -5.1696
#>   -0.5112  -3.1744   5.3173   2.9471  -1.4139 -11.4863  14.2095   2.5649
#>   -2.3681  -1.6252   0.0918   1.8547  16.1622   1.2585  -4.1964   4.7774
#>    1.9958  14.3396  -2.1460 -10.3246 -21.1537   7.7504   8.5280  11.2286
#>   -3.1101  -0.2383   9.2818   1.2624  -1.1347  13.5664 -13.8509  -5.6932
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]