Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.4044   0.3476  -0.5844   4.4491 -13.1267   4.8769   0.2586  -3.4340
#>    4.3336   3.4901   8.8769  -5.2411  13.5986  -0.9537  10.4550   6.9378
#>   -0.6293   4.8984   0.1107  -7.5186   0.8248   0.6588  12.2750  -9.3660
#>    5.3191   3.6766  -6.1173  -0.2109   2.1892   5.1547  -0.5877  -7.3460
#>    2.2622   4.6037   1.9133  -4.3078  14.0173  -3.0490  -6.4683  -2.8514
#>   -3.1088   1.2749   2.5930  -8.2264   2.6555   9.7468  -8.1336  12.7651
#>   -5.0103   0.4149  -5.4547  11.1377   9.9831  -9.8419  -9.1898  10.6131
#>   -7.1304   1.7136  -7.6914   1.6926  11.4590  -9.1120  -5.2334  -1.3226
#>   -2.0901   9.8701   1.3355   5.8086  -6.0542   6.9699  10.2352  -6.5950
#>   -1.7292  -2.0345   7.2206  -2.3417  -4.7041   9.4931   8.2242  -6.1336
#>    2.6956   3.5995   2.2505  18.7291 -10.4033  -3.2386   6.8634   4.5656
#>   -1.9539  -8.1069   0.8425   3.1780 -15.0443  -0.9737   8.3551   0.2959
#>   -2.6770  -8.3510   9.8766 -12.3362  12.6717   3.5533   5.6538  -1.9493
#>    1.3523   9.6398   2.1189 -14.4566 -17.6886  -6.3901 -15.5696   1.0623
#>   -0.5272   3.0825   3.3190  -6.0809   2.0461  -5.6255  -5.7709  14.5467
#>   -1.2830   5.8656  -2.5931 -12.4410   8.6349   0.9858 -10.9679  -0.6418
#>    3.1689  -2.0663 -11.4344  -8.5042   2.6463  -1.8659   2.6581  12.9046
#>    2.1027  -6.6132   2.9750   9.3532   6.2193  -7.2803 -13.9288  -6.8232
#>   -5.2012  15.9987 -12.6805  -4.5565   8.7698  -7.8083  -0.2914  -0.0796
#>   -4.8239  -6.1522   5.2397   4.1430   0.7579  -3.1889   0.0896  -1.2341
#>   -4.8251   4.9896  -5.7842   0.9855  10.1511  -8.3616  12.7562   4.8717
#>   -0.6307   2.9645  -1.0163  -4.1056  18.5574  -0.8340   4.0299  -3.1867
#>   -2.3292  -7.1266   1.6607   6.2552  -7.1623   2.1389  -2.7956 -11.3979
#>   -4.7655  -3.3633   8.1501 -11.8817   6.6574  -6.6166   1.6569  -0.8370
#>    3.3077 -14.5442   4.4389  -0.9248 -15.2341   9.4876   2.0836 -13.2158
#>    1.7899 -12.3888   3.5291   6.5512 -13.2162   0.6098   1.5926 -11.6110
#>   -1.6308  -5.7191   6.5372  -1.9432  -3.7063   1.4746   5.7596   0.1312
#>   -2.5192   2.0896  -4.1435   8.7170  11.5813  -5.7054  19.5464   0.8627
#>   -9.0788   2.7977  15.6882  -0.4495   7.5794  -6.4330  -3.8531  -7.7669
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]