Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.5559   7.1585   1.8792  12.3581  11.0188   4.6861   3.2634  -3.8289
#>   -1.0544   6.0934   7.4159  -5.0656   2.8373  -3.9403  -1.2304  13.5808
#>   -2.3090   6.0453   5.7731   3.3488   9.9844 -15.8737   2.1315   7.1418
#>   -3.7418  -7.8323  -5.9067  -7.0539   0.5428  12.2242 -14.9440   9.8251
#>    2.4289 -11.9941   0.3072  -0.3567  -4.0007  19.0011 -13.7060   8.6805
#>    2.3855  -7.2169   1.5501   0.9030  12.7962   9.6587 -18.5593  24.1343
#>    4.5228  12.8513   6.3640  -2.2318   6.6892 -15.7800  -2.6603  -0.5676
#>   -1.3299   3.4611   2.1787  -1.0294  -4.8799  -5.2364 -10.5712  -6.2876
#>    5.7220  -5.3960  -1.2972   5.7044   4.1280  -3.2727  -3.6424  14.5891
#>    3.7197   5.7278   4.5078 -20.8281   3.7209  -3.7737 -11.7671   8.1401
#>   -0.6450   6.8351  -0.8296   6.6565   4.8143   0.1678   3.9602  -1.7544
#>   -5.7685 -11.2398  -6.6906 -11.9987  -3.7946  -9.7765   2.6513   3.0499
#>   -2.0518  -6.5254   0.6295   0.6343  -4.0361 -11.2486   7.2208  -6.3308
#>   -7.5858   2.3962  -6.4642  -4.2143   3.8951  -9.6143   6.6520   4.7249
#>  -11.3295   1.1142   3.7378  14.2830   9.1376  -4.4867  -0.6174  14.2358
#>   -2.3860   2.7669   8.0073  -2.1703   8.8258  -1.5032   0.9237  -3.0506
#>   -1.0633   7.2200  -3.8992   0.0641  -6.7580   5.5831  -6.2437  10.0449
#>    3.6479 -10.0338   1.0872  -2.6731  -8.1700   5.9548   3.2593   3.1711
#>    3.8596 -11.0571 -15.9480 -11.1521  -4.9252  11.1292 -16.8823  11.8437
#>    8.4419   7.0595   1.0216  -1.1323  -4.4174  -1.9511  -6.0979  13.6870
#>    2.6152 -10.0953   3.8594   6.5306  -8.0027   5.3439  -4.1668   2.8858
#>   -6.9036  -4.3115  -1.0631  12.9459  -4.9309  15.4269  -1.5913  11.6078
#>   -0.9521  -5.1458   4.8908  14.5121   2.1616  -4.0565 -17.9282 -27.6609
#>    5.3129   7.1276   6.7950  18.4528   9.0892   4.5224   4.0355  -4.5651
#>    0.0603 -16.0220  15.0755 -11.5300   3.0959  15.7564  -8.5962  15.2135
#>   -1.8949  -1.7273   4.0716  -6.8101   9.2479 -15.3274   8.2678  -0.3470
#>   -3.3096   6.4002   9.0199   5.1996  -3.5036   1.6587  -3.3670   3.2291
#>    0.6177   2.2194   8.0530   3.5027   3.3106  -6.0620  -3.4582   5.7113
#>    3.2918   6.2504  11.4960  -7.1244   1.8805  11.3691  -5.1050  19.2318
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]