Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.2058  -1.4058   0.8320   3.2724   7.2165   6.9726 -19.3715   0.9439
#>   -3.4357  14.9220   3.0536  -8.8062   1.1997 -19.4858  -6.8326  -7.8271
#>   -9.6209   9.2336  -0.8870  -0.5284 -13.4202   4.2259  -2.1658  20.0644
#>   -5.3865  -5.2919  -3.7713  -4.4190   4.4306  -3.3762  -4.8797  18.7673
#>    2.7669  -0.3730   0.7915   7.8245 -13.3509  -4.9115  -4.4399   3.8730
#>    4.9730  -1.5397  -1.3620  17.0498  22.6374   4.3881   9.4733   1.5311
#>   -7.4313 -12.1730   4.6241   6.5178  22.5733  16.3235  -6.9757 -12.1941
#>   10.0300   3.6555  12.5640   4.2309 -20.4548  -5.1870  -8.1278   6.3481
#>   -4.9116   1.3185  -1.8063 -10.8085 -13.6772  -0.0821  11.1788  14.8861
#>    9.7120  -1.5788 -13.5484  -7.0227   6.1157  -1.4295   6.3298   0.6547
#>    4.2224  -1.6318   2.4347  -1.4193  -6.9794  -8.0829   4.0112  12.6260
#>   -0.3585   1.2762  -0.1895  -7.0128 -14.0892  -0.2574   3.9538  15.3102
#>   -0.1686  -2.1482   1.5940  -3.8195   7.7104   6.9816  -7.5834  -1.9437
#>   -1.5128  -0.8318   8.8484  -1.4961   6.6559   5.3621  19.2609  -4.1051
#>   -5.1976  -3.6466   6.8853   7.1419   1.1943 -10.4053 -13.4497 -24.3251
#>    0.1750   1.3812   7.4418   3.3755   6.7820   7.5930   4.5165 -12.2749
#>    4.3254  -8.4924   0.4639   3.4357  -0.9358  11.6000   9.9161   0.1898
#>    0.3591   8.6668  12.2833   0.2095   4.5169   5.2565  12.8619  13.9618
#>   -1.9973 -10.8156  -0.8267  -6.6190   5.2304  -5.4802   8.8069  -9.8624
#>   -1.5654   1.7938  -0.5882  -5.3059   0.6262 -10.7605   8.2705   7.0521
#>    0.5542  -7.3129 -11.3550  -2.0262  -6.3909  10.6901   1.9520 -11.3424
#>    9.3318  -5.4100 -16.4387  -1.9448  10.4323   7.1327   8.8031  -9.4973
#>   -5.9075  -4.0161  -3.8217  -7.8520   8.1103  -9.2553  -6.3744 -11.1753
#>   -5.2959  -5.6805  -5.7093 -26.6473  -0.5021   7.4834 -13.6479  15.1978
#>    8.8487  -3.2581 -13.2468 -11.3317   4.2258   6.5531  -2.0823   0.5039
#>  -10.8995 -10.2470  -5.5749  -3.4872   2.6083  17.3368   9.1688  -0.5056
#>  -10.7222  -7.1851   0.7264  -0.2774  -7.3033   4.2842 -10.8675   0.2471
#>   -3.9357  13.8359  -4.2193   1.3043 -12.4592   1.6835   4.6738  -0.9822
#>   -2.8230  -1.2140  11.2043 -14.2871  -7.3745  -4.5586  -3.0150  -0.3154
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]