Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.6390 -10.7645  -7.4875  20.5327  -2.8286   5.5712 -14.4051  12.1011
#>    0.7434   1.5723  -0.1057  -8.5181  -1.9664   7.6934  -0.7416  -8.4610
#>    2.2682  -2.3604   2.1127  -2.9451  13.0123   9.4526 -10.6504  10.4951
#>    2.3691   6.5141 -14.7239 -11.4809   6.6961   9.8274   4.7058  -3.2546
#>   -0.7589  -3.1062   0.1074  13.5842  -2.0580  -1.6564   0.4391  -7.3968
#>  -10.6479 -12.3597  12.6256   6.0064  -2.3068   9.1149 -13.2651  -1.9664
#>    4.5201  11.0123  -2.2530 -14.0533  -7.7925  -0.4955   7.7405   8.1025
#>    4.9158  -0.8539   5.7700 -15.6187   5.1932   2.2297   4.2517  -3.2179
#>   -3.5345  -7.3580 -10.7018  -3.0171   3.2087   5.0256   0.7169 -12.6269
#>    7.4775  -1.6178  -8.3992  -5.1860  -4.1079  22.5569 -14.0274  11.1167
#>    4.3528  -0.3706 -19.0935  -4.2923  17.5356  -9.2276  -6.3668  -1.1868
#>   -4.3619 -17.6141   5.6130   3.3886   4.2839  -7.0045  10.0305   2.9080
#>    4.4278  -5.4049  -7.0754  -9.0040   9.1342   7.8007   2.1885   4.1782
#>    7.1217   1.1836   5.9324 -13.7951   0.9728  -0.7324  15.7052  -6.6930
#>    2.6673   3.9088  -1.6946   5.6564   2.3341  -0.6956  -8.2095 -14.3566
#>   -0.5650   5.1507  10.5367   9.5807  -2.4523   5.9022  -2.7014  14.4385
#>    0.4614  -7.8577   3.3918  -3.9505  -0.0143 -10.7677  -5.0338   4.2265
#>   -3.6272  -5.3772   6.4281   3.8467   0.1806 -12.3297   7.6253  12.2465
#>   -6.3480   3.4440 -16.6252  -2.5795 -14.0441  29.8663  -1.3916   4.5314
#>    3.3561  -4.6450   2.5959   8.4830  -6.4917  12.8333   0.6743  -8.1547
#>    7.5284  -4.5041  12.5569   9.7867  12.2831   3.3959   4.9436   5.7820
#>    1.6314   4.2793   1.4793  -3.4308   0.4049  -0.6400 -15.5337   1.3096
#>   -8.5801   0.5100  -3.1838  10.2627  -1.0343  -1.8290   2.6082   2.3916
#>    1.0846  -1.5000   6.7272   2.2576   5.7543  -8.4162  -3.0774  10.1175
#>   -3.2073   4.8999  15.7982  -5.4358   8.7469  -6.2266   5.4424   6.0512
#>   -1.1605  -0.0377 -15.4166  11.6373  20.7018   5.5378   1.6090  -3.8257
#>    2.3991  -0.2668  -7.4759  -7.6701   9.7258 -20.4244  -2.7195  -0.3818
#>    2.9274  -3.1639  -6.7676   1.4385  -7.4312   4.6404  -1.5466   5.9746
#>   -3.5530   0.6665   2.4666   8.2960  -3.0810   4.2750   9.8443  -6.5549
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]