Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.2421   7.5251   6.1577   4.8258   3.8552   7.3055   0.4960   1.8050
#>   -2.7644  -4.9546  -5.0166  -2.8568  -2.9590 -11.3276  21.4641 -12.7148
#>    3.8941  -3.6146   0.2128  23.6709  -4.7264   1.1572   6.0001   6.6916
#>   -2.8003  -7.3204   4.2433  13.8282   1.5772 -11.2611  -4.4804  -7.4582
#>   -3.0524   5.5738   8.4039   7.8853  -7.1388  -0.4192  -3.8493   1.0811
#>   -2.7521  -1.9423  -4.9065  -4.7675   7.2016 -10.8166   3.8114  -1.9894
#>   -0.8779   3.1124   5.3083  -6.3459 -19.1613  -9.4636   7.2078   3.0121
#>   -2.0162   3.7179  -3.5193  -6.6074  -8.0182  -9.7636   2.3391   4.0114
#>    1.0166   5.9096  13.7973  -9.7465   1.5633  -1.9676  -6.2621 -13.7062
#>   -3.1059   4.9895  -1.2372   9.9760   2.8253   1.2166   3.4541 -12.4283
#>   -0.8801  -6.5155   2.2105  -1.2181  -1.5330  -0.8867  -0.8456   5.0968
#>    2.6973   2.3699   5.1621  -0.6771  -8.3512  -2.3504   5.8872  21.7186
#>    0.7847  -8.0896  -5.7920  -7.0167  -7.6083  -1.8059   1.6193   6.5602
#>   -1.1153  -0.9435  -4.5858  -0.4032  10.8829  -5.1368   0.8265   6.2172
#>   -1.2467   5.1530   4.4508   7.2386 -10.3033   6.3306  -9.2769  18.4348
#>    0.5591   6.0501  -6.2452   9.3022  -6.4818   6.0453  -7.0956  -1.1145
#>    0.7008  -4.8962  -9.4252  -1.9979  13.2981  -5.8280  -0.1689  -2.7927
#>   -0.0918   3.3704  -1.1294  -1.3437   3.1338  10.3825   9.3735   2.9676
#>   -0.8255  -5.0968   6.9852   5.9705   6.4244   0.5358  -8.0757   3.0388
#>   -9.3150  -5.6777   9.5190   2.9343  -6.3107  13.1541 -10.3837   4.2154
#>    3.3397   3.8288   7.9330   4.1586  -1.6217  -3.4357   1.5597  -5.1216
#>    0.5045  -0.5032  -0.0780   3.1423   4.9225   9.1066   0.4297  -7.6951
#>   -6.3968   0.8209   2.8157 -10.3661   2.6649   1.6324 -11.6597   6.2792
#>   -1.7277  -5.4051  -2.9638  -2.5243  -8.9297 -11.3201  -5.3337   9.7889
#>   -0.4056  -1.1734   5.8333 -12.5990   3.7332   0.8993  -6.0769   7.7044
#>   -1.4873   1.8056  -1.8987 -13.7492   2.3323  -2.6849  -2.0838  -7.2402
#>    1.4512   4.3583   0.8027  -9.3801  -7.9272  11.2879  -2.6828   4.4990
#>    4.5725   4.9532  -9.8857  -0.9634 -20.6468   0.0078  15.6241  -1.1887
#>   -1.5006 -10.7516  -2.2070  -1.5148   3.1901  -2.4946 -10.5385 -20.3669
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]