Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.9852  -3.7813   6.3663  -7.7116   1.5938   6.7957  -4.1424 -13.9084
#>   -2.7636  -1.7402 -10.4670   4.0248  -6.0866 -10.7319  -1.3090   6.5740
#>    5.6722   6.5441   1.3898  -2.9137  15.1686  -0.7728  11.4757  -5.2392
#>    1.2418  -8.8225  14.2104   3.0979  -3.6257 -12.9903   9.3970  -4.9913
#>   -6.2315  -0.9892   5.5683  14.2461  -2.3318  -7.5622   0.5288  -8.1180
#>   -3.2440  -2.7716  -0.9774   1.4467   1.6143  -9.2632  -8.3830  -2.0729
#>    5.7595  -7.0943  -7.6239  -8.0408   5.3698  -0.8993  -6.3553   1.2234
#>   -1.3912  -1.8449   0.2616  10.7541  11.8027   7.7458   4.1053 -11.6313
#>   -0.4602   9.3838  -5.1524  -0.1118   4.4702 -10.1715 -15.9623  11.5253
#>    0.8976  -5.7262  -5.1928  -4.4032   3.4608  -2.6220  -9.9519 -18.4698
#>   -5.0795   2.9459  -2.4302  -2.5354 -11.8920  -9.5457 -14.8597  -3.1382
#>   -0.4726  -6.2843  -3.5645   3.1355   0.5243   8.7476  -9.1830 -17.4979
#>    4.3279  -2.9729  -2.7362   1.5366  11.9501 -11.1750  19.0041   1.9435
#>   -0.1125  -8.0308   5.5900  -6.0555  -3.2509   7.9758  -0.6442  -2.3278
#>   -1.8035   1.7790   2.4759   4.0451  -0.7587  -2.1586  -2.9957   4.3636
#>   -0.9471   2.4705   4.1961  -0.4536 -14.6557  -8.3899  -4.4904   1.1842
#>   -1.0580   2.2734   0.3971  -4.8454  -1.5905  14.3973  -3.9257   4.3482
#>   -4.2576  -1.0052  -2.5492  -9.3777   3.0834   0.8948  -3.8794   2.2850
#>   -0.1615  -3.2366  -8.7840   1.5959  -1.2059   2.9002  -3.5041   4.1052
#>   -6.4504  -3.1492  -8.9297  12.6850   0.3268   6.3077 -10.8813 -11.1367
#>    1.9449   1.4182   3.6474   2.3816   6.3242  -6.7172  11.7264  11.1563
#>   -4.9827  -1.5595   6.3653   4.9076   0.4588  -1.2936  -0.4035  -9.8662
#>   -0.1051  -3.5298 -12.7401   7.8858 -12.6929   6.4402  13.5736  -7.9629
#>    0.7947   3.8420   3.9367   3.1154   2.8472  10.7414   9.2767   5.1366
#>    1.8041  -1.0211   1.6407 -17.6055   2.9649 -16.5924 -12.7033  -1.8966
#>   -0.7718   6.9210  13.2710   9.0294  12.4126  -3.8607   1.8829  -6.7207
#>    3.7694  -2.4198 -11.9579   3.4286  -1.4418 -11.2261   3.8016  -1.1001
#>   -1.0217  -1.3863   6.9602  11.0213  -3.1279   8.5089  10.7339  -7.6997
#>   -3.3204  -0.2315   7.8139   3.4130  -0.0461  -3.7193  -1.1158  -2.9552
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]