Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.3424  -0.8999  13.4123   7.1795  -2.7368  13.5310   5.9954   5.2939
#>    2.7901  -2.6051  -4.7244   4.2669  -4.7607  -2.9002  -4.7989  -0.6185
#>    4.1608  -1.2243   4.2497  -0.7551   6.2046  -8.0009   1.9357 -13.0368
#>   -0.4348 -11.2689   9.7529  -0.9769  -9.3932   3.0452   0.4964 -12.9862
#>    2.1243   1.7417   5.6165  11.4981  -0.5227   1.7499  10.0330 -10.9141
#>    7.2467  11.9656 -14.4620  -5.3785   0.4627  -7.8037 -14.8528  -4.9975
#>   -1.0623  10.3010  -1.9054  -1.6902  -1.2287  -4.8073  -9.1242  -6.4914
#>   -4.2645   4.1996  -8.9895   2.2592  -9.4881   8.3487   1.3352 -12.4473
#>    0.4127  -9.5625  24.4646   2.6170  -0.9898  -0.1163   2.0814 -17.2255
#>   11.1175   0.1214  -9.7537  -3.7949   5.8960   3.1131   4.6528  -7.7014
#>    2.9848   7.9920   3.1297  -5.5375   8.5131  11.0837  -8.6543  -2.4897
#>    8.2493   3.0546  10.3998   2.6484  18.9389  10.2001 -13.5338   4.9616
#>   -2.0229  -3.2798  -5.9970  -9.0471  -4.4478  -3.9079  -9.4791  -7.7786
#>   -1.2594   0.4708  10.8604   5.4733   9.1623   5.9625  -0.9484  -1.5879
#>    1.9720  -4.5166  -5.7518   4.3257   0.0554   6.5408   3.6668   2.6820
#>   -0.5087  -1.0066  -4.9128   5.6026  -5.1022   2.4877  -1.1163  -1.7146
#>   -4.0245  -1.3991  -5.5605  -1.8590   2.1522  -1.8387  -9.3409   8.5676
#>    2.3187  -8.6400  11.8306  -5.3511  -7.1065   8.9649  -5.6700 -16.0721
#>   -2.4423  -1.5733  -8.5829  -9.0161  -7.9364  10.5765  -7.1822   6.8487
#>    2.8692   0.9590  -5.1631  -0.7684  11.8168 -11.0747   3.0879  -6.9774
#>  -12.2072  10.1307  -2.1788 -10.8984 -17.2758  -2.5668 -10.8066  -4.2723
#>   -2.8333  14.4297   0.8171   7.2336   3.0671   1.0009   4.2179  -4.2082
#>   12.2771   2.3109   1.7081   0.6423  -8.6192  -3.7766   5.9040  -9.5184
#>    3.7569  -2.1457  -9.2989   1.0554   1.7355  -2.5359  -5.3331   1.0288
#>    0.1076   1.2140  -3.9328   4.3827  11.3411  -9.3258  13.1060   6.3144
#>    4.3236   6.6096  -7.5252   9.0971  -9.8143  -6.9979  -5.8256  -0.8646
#>    2.9994 -14.1877  -3.9082  -7.9506   1.8168  12.1576   6.4458   1.3550
#>    4.4966   6.0726  15.6747  -6.8905  -3.1652  -6.7186   2.7402  -3.6851
#>   -3.8681   4.1424   0.4887   3.4998  -6.9036 -15.1969   3.2576 -19.2831
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]