Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.6666  -0.3854   5.8666   5.4178   8.5237  -8.8344 -10.2886   3.3098
#>    5.2576  -0.7962  12.1822   2.5629  -6.7161   3.2815  -1.6297   5.9422
#>   -1.4876  -2.6148  11.6038 -14.1496   9.2064  15.8410 -13.1042   0.9700
#>   -0.6584   5.2185  -1.9291 -10.2695   9.8685  -7.3950  -4.5207  10.4189
#>    3.4979  -6.0859   8.7233 -12.0701  -4.5369  20.1905   4.6717  -0.2980
#>    2.1181   3.9684   3.8758   3.2378   0.9753   3.6620   8.9139  -4.6628
#>   -5.9398   2.9905  -5.9986  -3.2727  -5.6438  10.2896   4.3516  -0.4350
#>   -4.3042   0.9988  -8.2173   6.3137   9.0304  -9.2059  -4.6636  -0.2675
#>   -0.5104  -5.5526   8.1053  19.7832  -2.8457  -8.6211   1.1937   9.2750
#>    6.1027  -6.6127  -2.1475   2.8107   8.3080  -3.9574  -6.4665   0.1285
#>   -3.5251  -4.8323   9.2059  -8.6830   5.9999   0.9534   0.4523  -3.4143
#>    4.4964   2.3950  -4.3222  11.0079   7.4032  15.9568 -11.1892   9.5484
#>   -3.7547  -8.7800   4.6426   1.5997 -24.1075   1.9618   4.8613  -8.7618
#>   -3.1438   0.3910   7.8750   0.9831  -4.4915   3.6995  14.6882 -10.7156
#>    0.7392  -1.8874  -0.5850  -2.3070  -6.9319 -17.7195   8.7827  -4.2444
#>    5.2091  -2.1978   5.5903   6.8547  -7.6417  -5.2386   7.0545  16.1334
#>   -4.6813  -9.0065   4.3955   0.2687  -4.6677   2.4410  12.9661  -4.5440
#>   -0.0136  -0.9196   0.8183  -0.2603   2.4341  -6.0306  10.1088  15.6075
#>   -1.0437   8.0342   1.1214  -4.0847 -12.5788  -1.5937   1.7619   3.0664
#>    0.2212   5.8689   5.1566  -6.3872  -1.2758 -12.3218  11.3282  -7.9207
#>   -6.0928   3.5795  -2.7770 -10.3509  -1.7145 -18.1562  -5.7720  22.4943
#>   -2.1520  -0.9672  -3.8150  10.2002  -3.3151  14.4372   1.2847  -4.4304
#>   -1.2295  11.1256   3.2616  -5.8075  -1.3699  -8.5525   5.2046 -16.0032
#>    3.3359  -3.0856  -4.3016  -4.4437  -6.3264   2.0812   9.4581   0.9423
#>   -0.5814  -0.2384   5.5252   0.8575   6.8546  -4.4914   0.2513  10.0278
#>   -0.5557   4.8021  -3.2581   8.8060 -13.6710 -10.5010   2.2378  -1.0791
#>   -3.8846  -3.8498  -5.7175  -6.5678   3.6085   6.1893  -6.1210  -0.5948
#>   -3.9186   3.0942   5.5377   0.3978  29.5519  -1.8296  -3.3762 -11.4588
#>   -8.0709   2.9739 -14.0136  19.3410  -0.0719  -6.6088  14.3320   2.1333
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]