Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -3.5676e+00 -2.7837e+00  2.8197e+00 -2.6721e+00  3.2047e+00  3.0764e+00
#>  -3.9267e+00 -7.2893e+00 -3.5184e-01 -1.8891e+01  1.0923e+01 -1.7514e+00
#>  -6.2972e-01 -2.6519e+00  9.0780e+00 -1.6336e+01  7.3273e+00  7.5757e+00
#>   5.6947e-01  4.3709e+00 -2.3431e+00  8.3495e+00 -1.2409e+01 -8.9173e+00
#>   1.4390e+00 -4.3245e+00 -5.5403e+00 -1.4855e+00 -6.5182e+00  9.7672e-01
#>   5.1783e+00  2.9574e+00  1.3883e+01  5.1756e+00  1.4239e+00 -3.8933e-01
#>  -4.9283e+00 -8.8427e+00 -1.1069e+00  1.2327e+01 -6.8041e+00  5.4947e+00
#>  -1.6816e+00 -5.2655e+00  8.8690e+00  8.0022e+00  9.0769e+00  1.1435e+01
#>   7.5971e-01  2.4534e-01  3.3013e+00  1.6324e+01  1.5256e+01  7.4838e+00
#>   5.4213e+00 -4.8064e+00 -1.6552e+01 -7.5661e+00  1.9537e+00  1.2297e+01
#>  -5.7201e+00  5.5959e+00 -2.3884e+00  1.4895e+01 -3.8193e+00  1.2055e+01
#>   1.5365e+00 -1.2520e+00  8.0040e+00  5.4024e-01  1.7687e+01  1.0833e+01
#>   2.6282e+00  7.1848e+00  8.7981e+00  6.1550e+00  6.4276e+00 -2.6290e+00
#>  -4.3818e+00 -2.1897e+00 -3.9500e+00  1.1477e+01  1.4359e+01  1.1844e+01
#>  -1.9532e+00 -3.4190e+00  4.6411e+00 -5.2250e+00 -2.2407e+00  5.7611e+00
#>   1.1504e+00  3.4501e+00 -1.4922e+00 -8.4527e+00  6.4033e+00  2.2899e+00
#>   4.5785e-01  1.3460e+00 -7.5801e+00  8.3063e+00 -1.7716e+00  1.5621e+01
#>  -3.1716e+00 -6.2479e-01 -1.3363e+01  3.2192e+00 -6.6708e+00 -6.0660e+00
#>  -6.7636e+00 -7.9581e+00 -5.9063e+00  5.2107e+00  2.8864e+00  5.4657e+00
#>   1.0085e+00  2.9028e+00  6.4852e+00 -7.7908e+00 -2.9446e+00 -1.1697e+01
#>  -7.1125e-01 -6.5600e+00 -3.3593e+00 -8.9396e-01 -8.7771e+00 -9.7001e+00
#>  -4.7625e-01  5.5462e+00 -6.8946e+00  1.8921e+01 -1.2170e-01 -7.5245e+00
#>  -7.4742e-01 -7.1992e+00  1.3587e+01 -6.5071e+00  7.8519e+00  1.0070e+01
#>   1.7830e+00  5.2972e-02  8.1569e+00 -3.9362e+00 -2.5652e+00  1.2189e+00
#>   5.9228e+00  1.2563e-02 -1.3541e+00  5.8625e-01  9.3019e+00  5.9698e+00
#>   2.1143e+00  4.2631e-01  1.4683e+00 -4.2360e+00 -2.9379e+00  1.5014e+01
#>   2.1599e+00 -6.9365e+00  5.7573e+00 -1.7192e+01  7.9915e+00  2.0566e+01
#>   4.4680e+00 -3.5449e+00 -3.0314e+00  2.7848e+00 -6.0641e+00 -4.0046e+00
#>   1.4025e+00 -1.5995e+00  1.3688e+01 -1.3267e+01  2.1861e+00  4.2200e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]