Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.2900   5.6345  -0.2858   3.8828   5.4087  -5.8767   9.3669   4.1873
#>   -6.6799   1.5450   7.0192   3.0290  -3.6355   7.5915  -7.4347  -0.1176
#>   -0.3108   3.8599   0.4984  -1.4869   1.5669   7.9130  15.8244   0.7433
#>   -3.6481   0.8184  -1.9314  -7.2932   2.9590   1.7102  11.4961  12.2672
#>   -2.8559  -0.0942   6.4019   0.6285  -3.9263  -9.2918   2.0271  18.7645
#>   -3.1873   4.9215  -0.5557  -9.7974   1.9194   5.2689  -2.4958  -6.2664
#>    1.6243   0.5279  -2.3315   0.4088   5.7417 -10.0013 -13.8263  -1.7359
#>    3.2932   0.7648  -7.3056   0.2058  -8.0531   1.1583   0.7673  21.1112
#>    2.5986   3.7305  -2.2061  -3.3104   2.3218 -12.7932  -9.4452 -10.7865
#>    0.5541   4.0441  -4.7874  -9.9023  13.0820  17.7769  -8.8530   3.1290
#>    3.9713   3.4272 -15.7724   3.7127   5.9933  13.7576  13.7204  16.7033
#>   -3.5955  -3.7293 -14.0936  13.6005  17.5766  10.8910   6.5532  -3.1969
#>   -5.4827  -3.3614   2.4611  -0.1302   5.0736   9.4014 -10.2713   3.4404
#>   -0.1677   1.0410   0.3990   6.7127  -8.9110 -12.9709 -10.6592   2.9870
#>   -0.6085  -0.3535  -6.3511  -3.6282  11.7087  -2.6340  -3.1839   4.8217
#>    4.6200 -11.2337   1.2607   1.0946   9.9162   3.3455   7.6276  -6.3774
#>   -1.2365  -2.8681   3.3996  -1.5506  -3.0184   3.4948  -0.1529 -19.0003
#>   -0.0748  -4.5505  -0.4606   9.5521  -1.6640  -0.6759   5.3764   5.7965
#>    1.2768   7.2942  -6.4273 -15.6077   5.8035  -9.0331  -9.5284  -9.8212
#>    3.1306   9.8381  -1.4951   3.3750  20.8742   2.7754  -4.3763   5.0658
#>    0.3401  -3.9286  -1.9547  -7.8818   8.2793 -12.2825   5.0184   0.2345
#>   -1.7352  -6.7164   5.7687  -4.5676  -5.5665  12.8623  12.3822  -2.5704
#>   -1.8416   4.0372   4.8184   8.3639   3.2650   0.6685  10.9209  -0.1356
#>   -3.7925  -7.1966   6.5883   1.9890   8.1358   9.3880 -12.4906 -13.3367
#>   -1.0607   0.7413  -0.4205   7.1389   4.8261  -1.7801  14.9662 -15.8776
#>    3.1781   5.2981  -0.2806   5.2534   5.1168  -7.0428   1.0558  -4.1197
#>    3.1358  -1.1372   5.2695   0.1932  -6.4635  10.1510  -8.3219 -17.2889
#>    1.8091  -6.2443   0.5496   8.1755  -0.2620  10.4172   3.9176  -0.1896
#>   -3.7308  -0.8661   7.5791   2.4327  -8.7287 -17.7823  -0.7461 -12.9157
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]