Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.6080   5.1721  -6.0319  15.5741   8.0012 -12.4975  -2.7526  12.5928
#>  -12.7762  -3.1290  -3.0826   1.8038   3.6881  -4.2985  10.9288  -0.6844
#>   -3.7018  -5.5777   2.7282 -11.5654   0.7992  12.0353  -9.0451  -2.3299
#>    1.2362   6.9098  -2.9080   4.7955   0.1958  -0.2469  -0.4651  -4.4784
#>   -1.1436  -3.2827   0.8782  -9.1723  18.6944  -0.5787  11.0006   6.0949
#>   -1.2241  -1.8210  -5.5589  -0.7362  -7.1908  -7.4844   5.2792  -2.8797
#>   -2.7943   3.3771  -0.5953 -14.6293  10.8578   4.1663   8.1946   5.4348
#>   -0.1703   1.4961   8.1164   6.8318   5.9395 -15.9833  -4.5578  -9.8090
#>   -0.9688  -0.2455 -10.3644  -2.7679 -10.5598  14.1010   1.5783   3.4653
#>   -0.5212  -0.7774  -6.3194   8.9548  -3.4797   2.0802  -1.9292   2.1203
#>    1.0013   0.6858  -1.4871  -2.4038   4.4235  -6.3275   0.8686  21.2741
#>    0.9548   1.4239   2.9861   1.2361   1.8991  -1.9746   0.7989   8.2438
#>   -3.0257   6.3509   6.3787  -0.6775  12.8123  -8.8368   7.1653  -7.4538
#>    3.9711   4.6221  -0.7520 -15.5629   1.1936  19.5263  -3.1120  14.5616
#>   -0.3155   0.4766  -4.8687  -8.6736   4.7070  21.2076   1.1288   0.8656
#>   -0.9923   3.2639   2.9944   9.0801 -12.3085  -7.6014  -1.1456 -19.7339
#>    0.3008   7.4041   4.3929 -23.0364   9.8815   4.7139   6.0129   0.9736
#>   -7.2000   7.3224   3.8484   6.9085  -3.7496 -21.3168  14.7504  -7.7083
#>   -2.4271   1.8642  11.3259   0.8405  -9.1372  -4.1183  -4.8509  12.0025
#>    0.4045  10.0023   2.5885   0.5294  -9.1137 -11.4454   2.8684   6.1242
#>    3.5850  -7.4417   1.2648  -3.6674   9.1964  -4.8449  11.2576  -8.9955
#>   -5.0330   1.4303  -4.1611  12.1441   6.8799  -6.2137 -11.6145   6.3118
#>    9.9849  -5.1588   8.9320  -0.4345   8.8603  -1.3249 -14.0010  10.9819
#>    9.1144   0.3657  -2.0819  11.1507  -0.7373   2.9897   1.3781   1.1034
#>    2.7483  -4.1359  -4.2070  12.1308   1.3929   7.0402  -9.9770 -10.8140
#>   -4.1577   4.6079 -10.1547  22.0859 -11.5000  12.0156   9.0603 -37.6874
#>    0.4903  -5.4597   2.2856  -1.7711 -12.9033  10.1256   5.4120 -13.6220
#>    6.4548   1.9615   7.7983  10.1252  -3.2986  11.6578  -8.8018   4.3423
#>    5.1896  -5.8618  -2.4687  -1.7045 -11.0935  21.3552 -22.0360  -0.0578
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]