Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.3451   5.6344  -3.8803  -6.1248  -7.9329  -1.4407   4.0353   2.0156
#>    1.4527  -3.2745   5.2587   0.8740   7.7474 -13.5075  -7.8538  -0.1791
#>   -2.5482  13.0992  12.8235   8.3380  -9.7113  -0.3401 -17.0057   2.3657
#>    4.5446   4.0569  10.9503  -9.6049   7.9373  -5.7374   0.6126  -6.2088
#>    0.6178  11.2373   3.2064  -8.3658  -3.5492  -5.4310  -2.6177  -5.5293
#>    0.6449  -3.7290  -1.1860  -3.5989  15.6542  11.6334   2.1743  -5.3920
#>    4.4118   8.4749  -2.4640  -3.8102  -1.1317  -5.3378   0.4063  -3.0238
#>   -4.6432   4.2487 -10.2724  11.7922  12.7660  13.4318   3.0825  -9.3932
#>    1.5553  -3.2902   1.6154   0.8357   5.7157  -2.3789  -5.2956  -6.9974
#>   -3.7512   8.9784  -4.2081   8.2833  17.3872  14.4341   7.3745  -4.7563
#>    2.2171  -4.3635  -2.5129  -8.3423   3.9331   5.1160   7.7541   3.6526
#>   -5.5483  -3.2402  -6.2084  12.5878  -6.8025  -7.1668  -3.6361   0.9282
#>    4.7593   0.9928  10.0117  -2.7146 -10.0653   8.0189  11.9947  -7.7568
#>   -1.4037   4.7193  -2.5906   5.7416   6.1517   7.7403   3.6336  -0.6324
#>    1.1744  -4.3730  -5.8824  15.0032  -3.2823   3.1657  -4.6131   7.7865
#>    4.0365   1.4310   5.4951   6.8725  -2.4051  -4.5914 -16.2717  -0.3885
#>   -0.1143   2.2285   6.2308  -2.2024  -2.6081 -22.0697   0.7171  -5.9084
#>    4.7667  -4.0145   5.1780  -1.6811 -11.1073  -7.4327  -9.2491  10.0049
#>   -0.6643   1.1111  -3.4249  -0.2015 -16.5514  -2.7144  11.8262  -2.0469
#>   -2.1095   8.1380  -2.0388   6.2122  -6.2782  -4.2651  -8.8693  -6.5589
#>    0.3689  -8.1271  -6.3016  -8.8437  12.7688   3.5523   2.9663  -4.9475
#>   -2.3618  -6.4921   1.8002   6.9070  10.4672  -2.1950 -10.3798  -4.4005
#>   -1.9218  -4.9156  -7.0087  13.8640  -3.1991   7.4022  -5.7313  -8.5246
#>   -1.7256  11.1849   1.7376  -3.8294   6.1232  27.5843   8.2954   3.3467
#>    4.1404  -6.3377   8.1249  -0.8668   6.6636   1.4335  18.1716  -7.6333
#>   -4.6277   1.3974 -13.0520 -12.7928  -7.2391   1.3211   6.6353   0.8209
#>    3.3706  10.2918  -4.7954   5.3550  -9.0186   3.5210  13.1756   4.5359
#>   -0.3949  -8.1999  -0.2813  -0.9196   2.6301  -5.8687  -6.7661  12.3226
#>   -3.1604  -9.7500   4.4409  -2.9501  -9.6317   1.5290 -17.0951   7.2681
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]