Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -8.6342e+00 -4.6861e+00 -7.1518e+00 -1.8823e+00 -3.0488e+00  5.1271e+00
#>  -2.6961e+00  3.2646e+00  1.2454e+01 -7.2212e+00  1.7948e+00 -1.0944e+00
#>  -3.2235e+00  8.5070e+00  3.6299e+00  1.1301e+01  7.4807e+00 -1.1649e+00
#>  -1.9862e+00  1.7906e+00  3.2349e+00  1.0223e+01  1.1650e+00 -1.5503e+00
#>  -1.5510e+00  4.1361e+00 -3.5376e+00 -2.3790e+00 -1.4024e+01  7.5363e+00
#>   2.0015e+00 -1.3626e+00  6.7247e+00  5.8513e+00  4.6596e+00 -3.2649e+00
#>  -6.3951e+00  4.5111e+00  1.7683e+00 -2.1692e+00 -9.3686e+00 -1.2691e+01
#>  -3.7667e+00  5.5481e+00  1.4330e+00  4.2787e+00  1.3067e+01 -2.1664e+00
#>   1.8621e+00  1.1044e+01  4.1838e+00  1.3248e+01 -9.7205e+00 -1.0036e+01
#>  -5.6621e-01 -5.2329e+00  2.5314e+00 -9.3344e+00  2.4160e+00 -1.8481e+01
#>  -5.0238e+00  2.3725e+00 -2.1331e+00  3.3905e+00 -2.8117e+00 -2.0183e+00
#>   5.5972e+00  1.8297e+00 -1.1536e+01  9.1827e-01  5.3962e+00 -3.5148e+00
#>  -4.0478e+00 -7.7416e+00  4.1637e+00  1.0044e+01  1.4480e+01  1.5408e+00
#>  -9.5392e-02 -2.5531e+00  5.2606e+00  1.4495e+01 -1.1178e+01  2.9354e+00
#>   4.6927e+00 -6.1736e+00  4.9720e+00  1.6463e+01 -4.3886e-01  1.0566e+00
#>  -7.4870e+00  2.8578e-02 -4.1364e+00 -1.1751e+01 -2.7827e+00  7.4571e+00
#>  -3.2491e+00 -1.8585e+00  7.9733e+00  1.8001e+01  6.0319e+00  5.8632e-01
#>  -2.3485e+00  2.2484e+00  1.1488e+01 -5.6398e+00  1.0807e+00 -3.5950e+00
#>  -3.8396e+00 -4.5292e+00 -1.1271e+01  1.3905e+01  2.2094e+00 -1.6920e+01
#>  -2.7128e+00 -4.8890e+00  5.6812e+00  5.5739e+00  1.0211e+01 -7.6566e+00
#>   1.3769e+00  7.3135e+00  1.4650e+00 -1.7458e+01 -3.2520e+00  1.2868e+01
#>   1.1217e+01  2.6079e+00  3.0172e+00  3.6677e+00 -5.4855e+00  4.1308e+00
#>  -5.4142e+00  1.2036e+01  4.9182e+00 -1.1129e+00  9.7993e+00 -6.6789e+00
#>   3.6418e-01 -4.4026e+00 -8.5781e+00 -6.8385e+00  7.0802e+00  2.0237e+00
#>   3.7105e+00 -6.1891e+00  5.4063e+00  1.1450e+01 -4.5279e+00 -9.7035e+00
#>  -3.8720e+00  2.3830e+00  1.1752e+01  2.4519e-01 -8.6045e+00 -1.8664e+01
#>   2.3981e+00  4.6991e+00  1.0871e+01  8.6572e+00 -1.8120e+01 -1.6576e+01
#>  -9.7232e-01 -1.2518e+00 -1.0902e+00  2.9668e+00 -1.4117e+01  1.8240e+01
#>  -2.9527e+00 -1.1025e+01  1.2678e+01  2.9792e+00 -8.8076e-01 -1.1551e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]