Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.8709   0.2562   1.4639   1.6310   6.8802   5.7338  -1.4880  -1.2522
#>    5.2199   0.4887   7.2537   3.7812  25.1425 -14.3043   0.6895  -3.0590
#>   -2.5769   5.0610   6.8833   8.8293  -9.3168   9.3761   9.6640  13.5071
#>    4.3840  -8.3897  -4.1545   5.2346  -8.2820  -3.7185  -4.4582  -7.9534
#>    5.0424   1.2156  -4.2056  -5.6931   0.4818  -0.7489   1.5715  -7.3861
#>    3.0241  -9.4977   3.8438   0.2034   2.4907   4.9713   4.7819 -13.2250
#>    7.4618   6.6894  -1.0371   3.7848  -1.0629  -1.5415   5.0552   3.3140
#>   -7.8679   7.7913  -9.1959   2.2590   8.5713   8.1011  -0.8775   0.6996
#>   -0.6329   7.3151   2.8554   7.3878 -11.1956 -18.9896   1.3351  17.4314
#>    4.3626   0.8597  -8.4332  -6.3429   0.6908   4.2820  -1.9746   0.7098
#>    2.2532 -17.8677 -10.5552  10.4942   9.0181  -4.9074 -12.4352  -2.9569
#>   -9.7145 -12.3130   4.6171   3.1377   1.2093  -4.7561   4.0671 -18.3286
#>   -4.2563   0.7237  -0.7722   6.9629 -10.6218  -3.4203   1.3744 -13.1311
#>   -2.0356   7.2949  -3.1777  10.6543  -6.4070  -6.3864  -5.8651  12.5642
#>   -7.0080  -3.9778   5.0880  -8.6629   0.7105  -4.3513  -0.5791 -11.4885
#>    0.1199  -2.5883  -9.5327  -8.4484   5.6503  13.5716   7.9171   2.9765
#>   -5.1623  -3.7144  -7.6292  18.6489  -9.8661 -10.5922 -13.3837   2.2354
#>    2.4271   5.3381  -3.0032   4.0422  10.1637 -14.0435  -1.1308   5.2003
#>    4.9759   7.8299   0.3900 -10.2032  -9.5801   1.9108  -1.5206 -11.4628
#>   -0.8343  -0.7895   3.0744 -11.2969  23.0507 -15.5776   1.8294  -1.8390
#>   -6.7304  -0.4214  10.6611   2.1279  -8.8109  -5.5449   9.5882  -0.3302
#>   -6.1150   6.1929  -2.9600   2.0124   8.5659   5.2608 -18.2106   8.9976
#>   -3.2839 -16.3628  -1.7995   3.4166  -0.8412 -11.1134  -0.3508  -6.3246
#>    7.8937  -3.5132   4.3592 -12.1889   6.9600  25.5960   8.8939  -2.7422
#>   -6.6265  -9.0166   3.7704   3.9427  10.7359   0.2092   1.9211  -3.4324
#>   -4.6553   7.3987   6.8619   9.5856  -7.8534  -0.3653  -8.7262  -8.0281
#>   -5.4290  -7.5902   5.3121 -17.6671  -2.3639   2.9079  -0.6383  -1.1588
#>    5.9737   6.3007   5.9369   6.0082   2.2943  14.5166  13.6724   5.2294
#>   -2.5879   7.2428   9.3834   5.6455  -2.9561  -7.6086  -9.3310  -0.5157
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]