Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  2.5958e+00 -2.2592e+00 -7.7692e+00  3.0882e+00 -1.3300e+01 -2.6132e+00
#>   9.4564e+00  1.8918e+01  2.9998e+00 -2.3132e+00 -1.7636e+00 -2.3231e+01
#>  -1.1762e+01 -6.4678e+00 -1.3504e+01  8.0971e+00 -1.1474e+01  4.2161e+00
#>  -2.7236e-04  2.4878e+00  3.0187e+00 -8.0302e+00  1.4216e+01 -1.0975e+01
#>   6.0572e-01 -8.6363e+00  2.1980e+01 -1.1157e+01  2.4421e+00 -1.0098e+01
#>  -2.7757e+00 -1.3762e-01 -2.6407e+00  1.0896e+00 -1.4079e+00 -2.7522e+00
#>  -2.5155e+00  1.1213e+00  1.1492e+01  1.6352e+01  1.0723e+01  7.7940e+00
#>  -4.0670e+00  3.3761e+00  2.3258e+00  7.3708e+00  1.5936e+01 -5.4417e+00
#>  -2.5517e+00  2.8986e+00  5.1197e+00  1.4162e+01  4.9943e+00 -1.4372e+01
#>   4.3678e+00  7.4260e-02  1.0646e+01  3.5833e+00 -1.0605e+01  1.1545e+01
#>   2.1426e+00  9.5779e+00  4.9509e+00  7.6200e+00  7.9497e+00 -5.5894e+00
#>  -3.3118e+00  4.6746e+00  2.9502e+00 -8.7252e+00  1.3136e+01 -1.6118e+01
#>   8.7030e-01 -2.7683e+00 -7.0317e+00 -5.8995e+00 -3.7051e+00  8.3395e+00
#>  -1.3533e+00 -3.1545e+00  1.5802e+00 -7.3533e+00  6.0942e+00  1.3135e+01
#>   2.0110e+00  4.3130e+00  1.3280e+01  3.1485e+00  7.6521e-01  6.0194e+00
#>   8.4902e+00  6.5858e+00 -4.0373e-01  5.0368e+00  1.3199e+01 -2.0314e+01
#>  -1.1669e+01 -1.5702e+00 -1.7233e+01 -9.7079e+00  4.1811e+00  3.1942e-01
#>  -4.1658e+00  7.5787e+00  5.3509e+00 -1.1471e+00  1.8103e+00 -2.3286e+00
#>  -1.5772e+00 -1.1260e+00  2.3339e+00 -3.4453e+00  1.4333e+01 -5.3135e+00
#>  -6.5264e-01 -2.1829e+00 -1.7473e+01  9.8479e+00  1.2477e+01  1.3055e+01
#>  -5.8215e+00 -1.0903e+01  1.8425e+00  7.7106e+00  1.1612e+01  1.3515e+01
#>  -1.4080e+00 -4.4960e+00 -1.6961e+00 -8.8646e+00  2.2207e+01  1.3788e+00
#>   4.0486e-01  1.5164e+00 -9.2869e+00  1.7952e+00 -1.3762e+01  4.3502e+00
#>  -2.9675e-01 -1.6317e+00 -4.1567e+00 -1.0402e+01  1.4803e+01  5.4451e-01
#>   2.1896e+00  9.6079e-01 -8.2147e+00 -5.4722e+00  1.5581e+00  1.6545e+01
#>   4.9264e+00 -1.2363e+01 -6.0600e+00 -8.1642e+00 -1.0607e+01  7.8592e+00
#>  -6.1608e-01 -1.8703e+00 -7.7163e+00  7.1838e+00 -2.1052e+01  1.8689e+01
#>  -1.5170e+00 -3.1336e+00 -6.2709e+00 -2.2987e+00 -3.2094e+00  2.3329e+00
#>  -2.5872e+00 -2.1055e+00  2.6631e+00 -1.7975e+00  2.0579e-01 -1.8714e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]