Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.2620  -1.5541   1.3073  -1.6582  -8.6540 -12.5528   3.4901   1.9467
#>   -3.6687  14.9693   5.2128  -2.8389  -8.6013  -9.9948  -8.2089   1.1708
#>    2.6216   0.8672  -9.3027   1.5215  -3.5493  -4.8757 -10.3447  -4.3343
#>    2.1743  12.4496  -6.0443 -13.4190  -1.6846   1.0727  -7.3324  -1.2168
#>    2.1537   2.9683  -9.2452   1.3868  -1.6171  -2.0616  11.0507  -0.1434
#>    7.2226  -4.3812  -4.6396  -2.9164  -6.2085  -3.4617   1.5684   9.9503
#>   -3.8525  -3.2604 -11.5123  10.5714 -11.3689 -11.0120  14.5615   6.7753
#>    4.9356   6.8509   5.4574  -1.2010 -15.0201 -11.8177  -3.3885  -3.5603
#>    0.7574  -4.1903   0.7568 -16.0527   6.4207  20.5321  -9.2045  10.3704
#>    1.0623   0.7675  -6.4849  -2.1685   8.5138  -8.8019  -8.0084  10.8415
#>   -5.5874   2.6229  -4.7672   3.6517  -1.9539   5.1962   9.4311  -8.3756
#>   -0.7047  12.1362  -2.7392  -1.4281  -8.6337   3.7746   9.6313  -9.7009
#>   -0.1878   3.8831   2.4735  10.9941   8.4191   4.4241  -6.7076 -10.3123
#>    4.2502   8.3133 -14.8040  -4.9524 -13.3077 -21.6517  15.6753   0.6810
#>   -7.3745  17.3049   7.8889   2.7048 -11.5090   1.3381   0.6567  -3.6871
#>    1.7937   5.4722  -9.7362   1.2379  -7.5464  -2.0322  -7.9611   0.9502
#>   10.2496   0.3541   5.8707 -14.3813   0.1160   6.9267   3.9818  22.4115
#>   -5.4649  -1.1322  -8.8812  -1.0207   1.6542   7.9627  10.9212  -0.7210
#>    6.7558  -2.6663  12.5062  13.6728  -1.4918  -2.4636   2.5458  -8.7315
#>   -0.4164  13.1177  -1.2068  -0.7818  -9.2165   9.0433   4.6137 -11.6600
#>   -1.6562  -1.9577   0.0435  -9.9274   1.1351   2.2528  -0.1628  -8.5720
#>   -6.9058  -3.8411   1.1389 -19.6313   5.4995   5.9312  11.3710   7.9139
#>    3.0244   5.6697   0.8425 -13.6224  -2.4892   6.6476   4.1645   0.3614
#>   -2.9050   1.2220  -2.4370   8.6530  13.9697 -27.0964 -16.7588   2.4763
#>   -0.2931   3.5490   2.4061   0.7314   7.8659   8.0999  -3.7290 -12.9158
#>   -3.3470  -8.6368  -2.7580  -3.9741 -19.8142  16.5855  -4.4655 -10.7887
#>    5.9284 -15.7116  -3.7724  -9.0435   8.0083  -5.1889   9.6842   0.6971
#>    1.1546  -8.8733 -10.7731  -0.2254   6.1989  10.3481  -7.3499   9.1509
#>    5.9681   1.2167  -5.0070 -12.5552  -8.5429 -12.7980  -0.6059   9.1647
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]