Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   7.3898  -2.4845  -1.9680  -2.1214  10.6280  14.6545  -6.4387  20.6903
#>    1.9297  -1.8163   0.1849  13.2154 -14.4514  -2.0140 -10.2700   0.0133
#>    9.0849  12.1920   2.2697  -3.7250  -3.1962   6.6923  12.1440   0.5148
#>   -6.3040  -9.0539   2.8315   1.3492   0.9321  -3.4262  -0.6992  13.0335
#>    2.1813  -5.0571   2.5901   0.5625  -1.1917  16.1283  -1.2255   5.5028
#>   -5.1519  -3.2126   7.6178   6.2910   6.5994  -4.7811  -7.8297 -21.7300
#>   -0.5268   2.0726   1.0471 -21.0130   3.5529  -8.9743  -2.3706  -0.9077
#>    0.1999 -11.3458  -2.7690  -2.5788  -5.8515  -7.8429   9.2296 -22.0269
#>    0.7338   3.0915   1.1116   4.3833   9.4465  17.2852   6.9637  -1.3762
#>    2.8363  -3.9184 -12.0440   7.4480  -1.0067   2.6883   5.5731  -2.3436
#>    5.4733  -6.0568   3.5399  -8.0730   5.7900  -3.2304   0.8394   0.8631
#>   -4.0895  -8.8126  -3.9182  -9.6003  10.1827   3.8767   1.6259 -11.6277
#>    0.6574   5.5474   8.5855  13.3725 -11.6878 -10.3230  17.4300 -14.6339
#>   -1.9781   4.5186   8.2377  -6.4024  -7.0744   5.0548 -13.7808  -2.4783
#>   -2.4345  -1.2133   6.6283   9.1020   2.9592 -10.4520  -1.6440  -1.6631
#>    1.4255  -5.6585   6.4756 -12.2503   5.7042 -13.8513  -1.0199   4.7111
#>    4.6248  -4.8440   3.2227   4.1273   4.2760   4.0718   5.7248  -3.1597
#>    6.7909   6.3321  -4.3935  -8.1757  -4.2610  11.0448  -5.7579   9.5491
#>    7.2885  -5.2850  -7.5428 -19.3893 -15.6762  -8.6102   6.4085   2.6664
#>    1.0927   0.6378   0.9329   5.2747   8.3650   1.4011   5.7888 -11.9973
#>   -1.1315  -3.2017  -4.2915  -0.0718  11.4718  12.7172  -3.4165  -2.6957
#>   -0.6394  -2.7694  -5.1594   8.7364  10.4611   8.7993  -2.4057   2.9214
#>    3.5360   0.4615   5.9792   6.2872  -5.0532 -16.6711   6.5301  -6.6974
#>   -1.6222   7.9806 -10.0520  -4.5997  -0.4184  -7.1630  -3.9277  -6.7878
#>    0.9838  -4.2643  -2.6541   2.2892  -1.7683  -9.2618  -2.5203  -7.2682
#>   -3.6630  11.5593   0.5502   3.6128   3.4232   0.5256  -8.6180  -2.9541
#>    2.5918   0.2538   2.1914   7.8778  -4.9429   5.1389  -8.5070 -13.1886
#>   -5.0776   4.1760  -4.2237   4.9026   6.9966  10.1872   4.6806   1.1790
#>   -0.3933  -1.1311  -7.1734  -2.3055   7.2005 -11.7225 -18.4477  -4.0702
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]