Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  4.2510e+00 -4.3477e+00 -5.9874e+00 -3.4505e+00 -6.2301e-01 -5.4375e+00
#>   2.2411e+00  7.9563e+00 -1.4713e+01  1.9611e+01  1.7646e+01 -2.4458e+00
#>  -2.5937e+00 -9.2500e+00  3.1440e+00 -6.9800e+00  3.3399e+00  3.9373e-01
#>   8.9364e-01 -5.1540e-01 -1.4137e+00 -1.3738e+01  8.2875e+00  6.7039e+00
#>   1.6206e+00  4.6631e+00  3.0629e+00  8.0992e+00  3.1130e+00  1.1814e+01
#>   5.1995e+00 -1.2433e+00  5.5955e+00  7.9400e+00 -6.9713e+00  5.8102e+00
#>  -3.3789e+00  4.8373e+00  6.2396e+00  1.8732e+00 -1.0231e+01  6.8652e+00
#>  -1.0089e+01 -6.0983e-01 -5.6035e+00  3.3091e+00  1.2149e+01  2.3362e+01
#>  -7.8091e-02  6.7734e+00 -1.1921e+00 -9.4225e+00 -1.6812e+01 -8.3277e-01
#>   3.4805e+00 -1.3473e+00 -6.0788e+00  4.0171e+00  4.1189e-01 -1.4981e+01
#>  -1.5998e+00 -5.1829e+00 -1.4956e+01 -9.1831e+00  5.4547e-01  3.7527e+00
#>  -1.1041e+00 -8.8940e+00  8.9435e+00 -1.1677e+01  8.5695e+00  2.6622e+00
#>  -2.4174e+00  1.5551e+00  7.3956e+00  2.0417e-02  4.8516e+00 -8.0002e+00
#>   1.1583e+00 -3.8364e+00 -5.5881e-01  3.5494e+00  6.6072e+00 -6.0841e+00
#>   2.5040e+00 -6.4402e+00  5.4797e+00  1.0038e+00  4.7250e+00 -7.9699e+00
#>   5.5897e+00 -1.8181e-02 -1.7307e+00  1.1830e+01 -3.4203e-01 -3.4768e+00
#>  -7.0851e+00  5.0166e+00  1.2931e+00 -3.7820e-01 -3.9689e+00  1.6756e+00
#>   5.8983e-01  7.1047e+00  1.1950e+00  4.0269e+00 -3.5259e+00 -6.6840e+00
#>  -1.8380e+00  7.5502e+00  3.0148e+00  3.1256e+00  4.0547e+00  1.0088e+01
#>  -3.1708e+00 -2.9300e+00  2.0635e+00 -5.4248e-02 -1.8627e+01  9.8887e+00
#>   1.9918e+00  8.5455e+00  2.7508e+00 -8.4513e-01 -1.9456e+01 -4.0170e+00
#>   6.3392e-01  4.6814e+00 -9.0186e+00 -1.3740e+01  9.6684e+00  3.0685e+00
#>   9.2517e-01  4.2300e+00 -3.8852e+00  7.7422e+00  1.7551e+00  6.4916e+00
#>  -3.4284e+00  1.1312e+01  8.3695e+00 -1.6449e+01  1.1418e+01 -6.4309e-01
#>  -4.5206e+00  4.4604e+00 -5.2464e+00 -5.1266e+00 -4.6016e+00  1.6502e+00
#>  -2.1619e+00 -6.8393e+00 -5.8743e+00  1.3039e+01 -1.4549e+01 -1.3469e+01
#>  -1.1001e+00  5.6794e+00  1.1020e+01 -3.9671e+00 -8.8128e+00 -6.8887e+00
#>  -1.0701e+01 -6.0416e+00  5.2011e-01  5.1011e+00  3.7441e+00 -4.0340e+00
#>  -1.4853e+00  4.2471e+00  4.4995e+00 -1.3015e+01 -1.1429e+01 -1.3459e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]