Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.2783  -7.2449  -3.9380  -5.3757  -1.3346   0.7319  17.1295   9.4270
#>    3.8687   0.3550   3.9215   6.6207  11.6371  -3.8953 -12.9811   2.2872
#>    4.5505  -8.6211  -3.5104  16.5602  -6.2626  -6.2777  -1.0681   1.7654
#>   -0.3198  -2.0780  13.9050   0.1664  13.3431   5.7622  -9.3518  -7.3992
#>    2.9835  -4.5474  -4.0131  -3.8096  -0.7762  -1.2173  11.6312   5.7708
#>   -0.6420   2.2031   1.2846  11.4939 -13.6478  16.2049  -6.7550  -2.0056
#>   -3.4283   2.3056  -2.0306   2.5075  -2.9471   9.6880   6.4873   2.3758
#>   -5.6001  -1.6007  13.9880  11.2514 -21.9243   5.5649  16.6240  -2.4769
#>   -0.3690  -7.9339 -10.0922 -16.4867  -8.9476   6.9151   4.3433  -7.0748
#>    1.3736   3.4432  -1.4277  -5.2759  -2.8517 -27.4372   4.1821  10.6829
#>    6.2433  -3.1754   1.2484  11.0896  -5.1436   0.1904   2.2996  18.5835
#>   -4.2447  10.5087  -6.9053  -5.2828   4.9403  30.7037   3.4416  -5.2006
#>    2.9366   3.4935  -7.3075 -18.7721   9.0103   8.8252   2.9528 -19.4808
#>    4.2018  11.1089 -13.7913   7.5443   3.7615  -2.0557   1.7689  -2.0233
#>    2.6859   0.7971   7.7657  -7.7989   5.6374   5.2113   3.5601   3.5546
#>    4.9071  -6.0007 -19.1925   7.9147  -8.7520  11.1159  -4.1566  -1.1733
#>   -1.6305   0.2013   8.1848   5.0297  11.2377  -6.9329  11.2883   1.2666
#>    1.4693 -11.4036   4.1551   6.8717  -3.5366 -11.5536  -7.7038  16.0048
#>    3.8168  -3.4052   3.2702  -7.2942   2.7371 -10.6521  13.4181   2.5602
#>   -6.1770   1.2443   2.4304   2.0235 -10.3648  16.6703  -8.3376   1.4236
#>    1.8095  -9.9445  15.7442  -7.6067  -5.1457 -26.1041   0.1279   2.1984
#>   -0.0036   8.7372   5.8476 -10.3445 -27.0664   8.3912 -32.0924  10.7785
#>    2.1539   0.9022   4.9970  -7.8507   7.6514   1.3252   9.8412   4.7398
#>    4.6375   2.4690   5.9691   1.7583  -2.8554 -23.1758 -15.6790  15.1649
#>    1.0364 -16.2022   4.3442  -9.9092 -19.1735   3.4381   8.0059  -0.9712
#>   -0.7026   9.9149   8.6870   8.4442   4.4082   3.8047  -4.4556  -6.4965
#>   -0.5445  -4.0306  -2.7510  10.2493  -5.4829  14.7713  -4.5758  -4.2456
#>    3.1025  11.7637   5.2689  -4.2909 -12.1894   3.7632   4.5803   9.8500
#>    0.1399   5.4286 -14.5792 -11.3817  -2.3309  -5.8222 -16.7095  -9.0658
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]