Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.6956   4.8687 -15.5935  -0.4097  -2.7236  -2.6074  11.9783  11.3068
#>   -2.3060   7.6583   3.6856   1.5702  -8.7882  -5.9016  19.7703   3.2100
#>    1.4165   0.6111   1.8317   1.5781  19.5655 -10.4491   8.9647   7.5525
#>   -0.7791   4.2389   7.1009   3.1266   5.0096  15.5592 -18.0332  -8.9747
#>   -2.9372   0.1591  12.8350  -3.6519   5.0080   6.4611   8.5552  -2.6236
#>   -1.0179   0.9779  -2.2601 -11.2742  -6.5406   4.8271   1.2796  13.1115
#>   -3.5472  -6.6734  -3.8846  -1.8228   0.8209 -18.2204   0.2337   0.3115
#>   -0.5856   1.5025  -2.1698  -0.1372   1.6099  13.5843  -2.9311 -18.4333
#>    1.1094   4.1461   4.3948  -4.7396  10.3033  -9.4123  -8.3729  -5.8753
#>   -4.7690  -2.5800   6.5610   9.1487  -7.2891   1.4140 -14.8966  -1.2085
#>    3.4374   0.7965  -5.0706  -4.5341   9.2287   7.7011   0.7616   3.9162
#>   -5.2462   6.3598   1.0992   6.8705   1.1768   5.2455 -15.4302  -4.2492
#>    7.1931  -1.7675  -9.8044  14.0967   0.9959   2.5981  -9.3296   7.8604
#>    3.1084  -1.4543   1.1805 -11.4684  10.4187  -0.8647  -5.3166 -14.9463
#>   -5.9218  16.6095   2.7251   5.5819 -25.8425  17.9342   7.9615  -7.4665
#>    6.1679   0.7711  -4.2578  -3.0635  -2.2249   1.3431   7.9948 -11.6536
#>   -4.4247   2.9847  13.4060   5.5589  12.5517  -0.9735  -0.9895   4.0495
#>   -6.7833  -7.3756  -2.1541  -0.4820 -10.9821  14.3294  -1.1783  -9.2990
#>    3.1615   0.0302  -2.0800  -1.0891 -12.0585  20.7261   2.0019  -9.4776
#>   -3.5113   4.6455  -5.7907   8.0380  10.2272   9.7430  15.2731   2.2247
#>    1.3106   6.5670   8.5698 -14.5354   7.3628  -9.1580   1.0085  -4.5921
#>   -0.8059  -5.1053 -12.9732 -14.6121  14.0164   1.6834  14.2237  -4.7236
#>   -2.3096   5.9515   3.5643  12.9877  -8.4166 -16.9213  -4.3292   5.1660
#>    3.7918 -11.6706  -5.7545  13.1211  -1.9757 -14.2977   3.0301  -1.3763
#>    9.7496   6.9111 -16.6513  -0.3100  10.7074   0.7707  -9.3743  16.1997
#>   -9.1369   3.4322  -3.3434   5.4727  -2.5100  10.8166   1.9063 -15.5950
#>   -3.6916  -6.2534   2.4409   5.3215  -3.3091   8.7806  18.8651   4.5581
#>    6.8373  -7.2768   1.4855   0.5983   5.3300   8.5662  -5.8574  14.3351
#>   -3.4266  -9.5666   1.3723   2.9617   4.0878  -6.1277   0.3926   2.2879
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]