Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.8115   2.2349  -0.0214   6.0756   1.3913   8.7761  -1.1908   5.6461
#>   -6.0333  -6.6684  -9.3454   6.3885   0.7390   0.8873  -0.7448  10.9634
#>   -3.6137  -0.0686  -6.1895   0.6281   3.8656   6.9920  14.5680  -0.8502
#>   -2.4951  -2.2022  -5.1020  -0.3549  -0.8914   4.0418  -3.6507   2.4059
#>   -5.9046   0.3263  -0.3594  -3.8504  -8.8637  -9.1195   4.0265 -15.7326
#>   -0.2962  -4.5821  -3.4103   1.7672   2.6293  -1.4363   0.8268  -1.8779
#>   -3.1150  -1.6782   7.8295   3.1567  11.5828  -4.6585  -6.0042  -8.3150
#>   -0.2079  -2.4044  11.7208  -7.3103   8.0096  -3.4079   4.5758 -13.0278
#>    0.2965  -1.9558   0.6251  -1.6168  -6.8436   7.3008  -2.0629   8.2410
#>   -3.6611  -1.0853  -0.6529   1.6384  -0.1466  -1.3548 -17.7250   6.1867
#>   -2.2932  -4.7604  10.4790  -1.6942   0.6295  -1.2700 -12.0308   4.8470
#>   -0.5865   2.6289  -4.8464  -3.0038  -9.0615  -5.8440   1.9712  -8.9474
#>    1.5472  -2.8096  -9.6824   5.8388   1.4909  14.0926  13.3095 -15.9716
#>    6.3340  -8.6035  -7.8639   6.6698   9.5984   4.4203   9.3844 -11.2787
#>    2.2839   4.0051   6.3595   5.4781 -11.3676  11.0410  -7.9452   4.3184
#>   -4.9414  -1.3255  -3.9526   5.4606  -0.1756  -2.9342  -3.3871   6.5693
#>    4.2906   0.6060   3.7146   1.8299   6.2999 -22.4321   9.6698 -13.9056
#>    1.9161   1.8802   1.1212   4.3603  -0.9901  -6.1880   8.0598  12.3872
#>    2.6414  -0.1051  -3.6790   0.9326  -4.2497  -9.1232  11.7246  -7.7870
#>   -0.9460  -6.8372   1.8782   5.4102 -11.3042  17.2551 -17.3634  -2.7455
#>   -0.1574  -0.4721  -3.8736   7.2045  10.7730   5.9558  -5.5150  -5.3881
#>    1.6419   1.7210  14.2542  -5.8565   6.2499  -3.9261  12.7475 -16.9411
#>   -7.9782   3.1572  -0.5886   5.2288  -8.4329   4.5580 -23.1700  16.6198
#>    1.2595   6.1016   4.2294   0.1943   1.7756  11.8837   0.9128   0.5980
#>   -5.4303   1.5756  -3.5061  -0.8242   8.4263   9.3156   8.2625 -19.7124
#>   -5.4662   0.2427   4.0634   7.4611  13.9718  -0.4862  -0.5059  18.5297
#>    9.2661 -11.6628  -6.8196   3.5532  -6.7579  -0.5042   2.6221  -6.9812
#>   -0.0979   5.0604   0.6323  10.2698  -1.9292  10.9395  10.8861  17.2200
#>    1.0564  -7.5971  -5.4708   9.7854  -3.3543  -4.6503  -4.0061   5.7691
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]