Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   7.1902  -4.2384  -6.8612  -1.4895 -10.4800  19.3912 -14.5039   4.7190
#>    5.5082 -10.2830   0.0255  -0.0691  -7.7885  -9.7128  -1.6373   5.7670
#>    1.8280 -16.4233   6.7349  -0.8871  -9.2265  -9.8601   1.9603   7.8475
#>    4.3000   4.3940   6.8957  -3.7683   2.7335  -4.4912 -17.9945  -4.1384
#>   -0.3066  -1.7171  -6.6052   1.4522   4.3128  11.5515  -3.4462   9.3645
#>   -3.8030  -2.9145  -7.0871   1.2973 -18.5102   5.5586   2.3286  -1.1249
#>    0.6082  -0.4057   5.6733  -1.1174   6.0070  -1.0075   6.4426  10.8815
#>    3.5439  -2.4689   5.2585   0.7153   0.2628  -3.3469 -10.4678  -0.3284
#>   -3.1811  -0.4517  -3.5077  -3.8429  -9.6257  -0.1447   4.4568  -2.2818
#>    0.4590  -3.0389   3.1319 -13.8931  -3.4317   0.3426  -7.8219 -16.8504
#>   -6.1451  -1.7458  -5.0702  -1.2066  10.7499  -3.7216  18.2723  -5.1658
#>   -6.9925  -9.8995  -2.3827   6.5474  12.7225 -10.2829 -14.6320   7.4445
#>   -2.2270  -1.1518 -14.9951 -11.0841  -3.4531  -8.9159  -5.3802 -13.7952
#>    6.6844  -4.4812  -0.4180  12.8478   5.9127 -13.3042  11.7312  -5.9671
#>    4.4862   9.2654   0.8226   2.7220  -0.8814 -11.9164   3.7799  -4.5495
#>    9.0504   7.8280  -3.1677 -10.9377   6.5301   6.6051  -2.9323  19.9812
#>   -3.4281  -4.1584  -9.2815   9.5686  14.3087  12.0549   9.5495  10.6377
#>    0.1093   8.0729   6.8657   1.9782   1.1762   1.9403  10.8279  -4.7245
#>    5.1027  -0.6470   1.6992   7.3438  17.7485  12.5335  -1.5182   6.1231
#>   -1.6073  -3.5048   1.5263   9.8063 -10.2738 -10.3828   0.4126  -0.3989
#>   -2.9211   2.8626  -0.3088  14.4244  -0.2468 -10.8290   3.4499  -2.7622
#>   -0.2918  -6.0746   2.4929  -4.5316 -14.0659  -3.4907  26.7707  -3.5917
#>    3.1302   2.4621   4.5949   2.6358  -5.7472   1.7785  -4.2346 -14.4687
#>    4.3524  -6.1823   1.6304   3.2897  -9.8410  -4.9105  -0.1356  14.4486
#>    0.4962  -9.0968 -10.5387   0.9187   1.2597  -7.4495  15.2529  -7.0156
#>    4.8966   2.8159   4.4114  -4.0373  -8.7478  -4.2451  -6.0908  -0.9378
#>   -3.4960  -0.6899  -2.9133 -13.8674  -6.2538   9.9786  -0.8168   9.4613
#>   -1.7041  -0.6449   5.0084  -2.4926  -1.5202   3.3054 -16.5728   1.3568
#>    2.6760  -4.5550 -12.6681  -8.8449  15.0502  -0.2701 -17.9263  -8.9188
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]