Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.3016   1.7700 -10.8030   6.7142  21.9344   9.5027   4.4316  -1.9315
#>   -4.1222   2.1106  -3.2256  -1.2415   3.6064  -8.0887  -0.1894   3.7414
#>   -5.1666  -5.0083   2.1328  -1.2854  -5.6996  -2.6628  14.3558  10.3943
#>    3.7341   0.5295  -7.4815  11.0883  -4.8241  11.6925   2.3818   7.1568
#>    3.1849  -1.2695  -9.8042  18.5150   0.6368  15.5552  14.0779  11.4495
#>    2.4176  -6.7433  -1.3779   6.2697  -3.8637  -8.5074  -2.7939   2.5659
#>    1.3867   3.4524  -4.9672   7.5974   6.5245  -4.5798  -7.6735  -8.8197
#>    4.2815  -6.2087  13.0662  11.3965   2.3618   2.9382  -9.5834   1.0356
#>   -4.1487  -3.7875   8.9707 -11.2973  10.2405  -1.6850  -2.4736   0.3581
#>    2.3417  -2.7848   5.1305  11.4931  -7.3797  -6.3585   4.0390   6.2538
#>    1.0327  -1.0320   8.7099   3.0190   2.1384  -0.2362  -0.7861   7.2176
#>   -2.5581  -2.8760  -0.9045   2.1664 -13.6325  -3.7858   9.8178 -16.6506
#>    4.3049  -2.4302   4.7973  -0.0856   3.0812   9.8381   1.6417  -1.0050
#>   -6.9928   8.0893  -6.7788  -5.7669  18.2122  -9.3808   9.0375  -0.2547
#>   -2.3717  -9.0490  -0.2648   3.7305   4.3966  10.5494  -2.4866  -9.1694
#>   -3.2564  -5.6040  -0.4135   1.8049   8.6318 -12.9598 -17.1973  -1.1251
#>    2.7701   3.2773 -10.4012  14.1832  -3.0693 -21.6610   6.1517  -5.5618
#>    2.0380  -5.0658  -0.8172   9.6568  11.1150  -4.1658 -10.8914   8.3405
#>    2.4838  -1.7004  -3.8034   8.9228  -2.7981   6.1683 -17.0780   1.1753
#>   -0.1519  -0.4508  -9.0410 -10.1790  -2.4355  -6.3129   4.2978  15.8745
#>   -2.1804  -4.9735   7.7215  -4.0401   4.9080  -2.4948  -4.4441   6.7846
#>   -0.0302  -2.2978  -2.3171  -3.2072  -5.1452  11.7807  -2.3070  14.2241
#>   -3.0154   5.9255  -4.1714   1.7257  -6.1435   4.6606  -2.1654   0.8223
#>   -3.6594   1.7292  13.8270  -5.0814   2.6180   3.3988  -6.6161  -0.3972
#>    2.0214  -1.3182  -3.3607   0.4305  10.9545   3.3134  -3.6067   4.3397
#>    0.4954  -5.3916   0.9676  -4.7183  -8.5938 -16.1198  -7.9360 -13.7617
#>    2.0576  -9.1164   3.7041 -10.5478  -7.0029  -6.8146 -12.4259   0.1463
#>    2.6729  -3.3205 -10.7863   6.1142  -8.5504   7.1373  -1.3898  -0.0279
#>   -0.6042  -1.8052  -3.4579  -0.7208  -8.9686  10.2645  -4.5615  -0.3745
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]