Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.1648   0.6511  -3.8696   1.0516  -3.4251   8.5589  -8.7655  -0.7193
#>    4.2220   3.2033  -6.8519  -0.2563  -2.8693  16.5087   5.3591  -4.2724
#>    0.0751  -4.5923 -13.7749  -0.5195  -0.0773  -7.2011 -12.7707  -8.5927
#>    0.7910  -1.1261   6.2735  -5.5675  12.7673  -0.0339   7.4536  11.8393
#>   -3.0982  -6.0088 -11.8340 -14.6557  -0.1298 -15.3821  -2.0814  -0.2170
#>   -0.8262 -10.0205   2.1536   4.2886   8.3600   9.5733  -8.2543  -6.4737
#>   12.1944   5.6794  -3.9904 -12.1067   8.2415 -13.5783   6.6444  -2.6073
#>  -14.0674  -4.6971  -3.2905 -14.6155   0.4998  -6.5100  -2.2445   5.1002
#>   -2.6156   1.8534  15.0167 -11.1451 -14.6062  -0.0574   3.9497  22.7976
#>   -5.3872 -10.3601  -5.4171   3.5768   8.8421  14.5466  14.4085   1.3552
#>  -11.6125   4.0002 -11.6727  14.0915  17.8152   5.2169 -14.7570  19.2595
#>    1.6136   2.5344  19.0471   5.1700   2.2658  -5.3677 -11.6641  20.5677
#>    5.5598  15.7254  13.6157  -3.1656  -0.8567  -1.6031   7.3785  13.6663
#>    0.0603 -14.5396  12.6336 -19.9392  -1.2013   7.3884   5.0555  -1.6867
#>    0.0823 -15.3589  -5.7998 -14.4116 -10.4492   1.7436 -11.1468  10.0365
#>    3.6116  -6.3057  -9.2581  -7.3388 -13.8219  -8.3556  -3.7008  -1.7693
#>   11.9852   2.5586  -5.6732  12.5425   6.9618   5.1382  11.5247 -13.7397
#>   -7.2187  -8.5474   0.0217  -3.4423   8.8348  -2.0995  15.0909  -7.4510
#>   -6.4603   0.5591  -5.2344   9.3021  -8.4481 -19.1668   6.6774  -2.8232
#>    9.9722  11.4174  -9.1875  -3.0939  -1.4689  17.6936  -4.5090   3.0418
#>    9.5920  -0.8745   1.9121  -0.8611 -16.3454   5.1133  -3.4582  -2.7441
#>   -4.3820  -7.6693   3.6073  10.5652  15.3540  -6.7634  -0.3146   9.0035
#>   -9.3831   5.8770   9.4032   2.3070  15.5868   0.0457  -2.2698  19.1160
#>    1.4570  11.2949  -4.8892  -1.2794  -9.1057   1.6652   1.7916  -7.7607
#>    0.8697   6.4580  -3.6970   6.7926  -8.3627  11.1297   4.6777  23.7797
#>    8.7553  -4.0238  -3.3332 -11.7827   4.1669  -6.6496 -10.6363   8.1382
#>   10.7058  -0.0120   8.4496   6.7658 -13.9114   1.7716  14.9509   7.3951
#>   -6.6418  -4.3733  -0.6962   1.1255   9.8358 -10.4091  -1.2811   3.7386
#>    4.2744 -10.7465   6.2517 -12.5696   2.9644  -1.0260  -5.8955   0.6699
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]