Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  3.1173e+00 -6.0024e+00  1.4449e+01  1.0644e+01 -1.6712e+01 -1.3857e+01
#>   1.0396e+01  6.7444e+00  3.9494e+00  7.1182e+00 -8.4711e+00  9.8699e+00
#>  -4.8473e-01 -5.5560e+00  1.4557e+00 -7.4419e+00  4.7509e+00  1.0096e+01
#>  -7.4195e-01 -5.3199e+00 -4.9125e+00  9.4005e+00 -3.6736e+00 -4.4894e+00
#>  -3.1463e+00  1.1925e+00 -1.3855e+01 -1.3479e+01  9.5856e+00  3.9731e+00
#>   5.5852e+00  1.2690e+01 -8.1072e+00 -1.2694e+01  5.3283e+00  1.5609e+01
#>   2.9916e+00  3.2054e+00 -1.8956e+01  8.1787e+00  1.2133e+01  1.1796e+01
#>   4.4493e+00  5.8727e+00  1.3503e+01 -1.1933e+01 -4.5934e+00 -9.0441e+00
#>  -9.3687e-01 -5.3502e+00  1.7093e+01  3.0128e+00 -6.1053e+00 -1.1476e+01
#>  -1.0208e+01  2.6334e+00 -3.9783e+00  1.8961e+01 -4.3794e+00  8.4403e+00
#>  -8.0823e-01 -1.1015e+01 -5.8242e+00  1.1488e+00  1.7420e+01  9.6171e+00
#>  -7.4885e+00  4.4146e+00 -3.0624e+00 -6.9133e+00  1.2529e+01 -9.0778e+00
#>   4.9136e+00 -4.8246e+00 -4.5848e+00 -4.6168e+00  1.6090e+01 -1.5344e+01
#>  -8.4946e-01 -8.4450e+00 -1.0298e+00  3.7123e+00  1.3514e+00 -6.3718e+00
#>  -3.7995e+00  4.4979e+00 -1.4941e+01 -2.9779e+00 -1.3679e+00  1.5225e+01
#>  -8.2947e+00 -2.4852e+00  1.4190e-01  5.3268e+00 -7.4083e+00 -2.1856e+00
#>  -3.5376e+00 -1.0347e+01  5.2479e+00  4.7414e+00 -1.6894e+01  3.2094e+00
#>  -2.7281e+00 -3.9615e+00 -2.8645e+00  1.7094e+01  1.9825e+00 -1.8702e+01
#>   7.7102e-01 -1.0663e+01  7.6264e+00 -6.7383e+00  1.2197e+01 -1.8152e+00
#>   1.2881e+00  4.4213e+00  5.7875e+00  2.1701e+00 -1.2761e+01  1.3960e+01
#>   1.1306e+00 -1.5623e+00  8.8218e+00 -1.2401e+01  4.2394e+00  6.5970e+00
#>   2.7802e+00 -6.8587e+00  1.4639e+01 -3.6096e+00 -3.4631e+00 -1.1372e+01
#>   2.4473e+00  1.4630e+01  4.5562e+00 -6.3661e-01 -9.9495e+00  6.3231e+00
#>   1.6465e+00 -1.4867e+00 -7.4260e-01 -1.0928e+01  1.6384e+00 -1.4663e+00
#>  -1.1176e+01 -6.7904e+00  5.7621e-01  3.6838e+00 -1.0193e+01 -1.0677e+01
#>   1.0216e+01 -1.1143e+00 -1.0996e+01  6.7381e-01  1.0824e+00  7.1249e+00
#>   2.8397e-01 -5.8618e+00 -8.2264e+00 -7.3633e+00  8.0970e+00 -6.8072e+00
#>  -9.3754e-01 -1.5352e+00 -5.0819e-01 -7.0315e+00  4.5084e+00  1.6938e+01
#>  -5.5276e-01 -6.1329e+00 -1.5371e+00 -1.4730e+00 -7.2100e+00  1.9169e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]