Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.8900  -2.4387  -5.6081   5.4299   1.4312  -7.9818  -1.0467   5.2144
#>   -3.1542  -7.2749  -7.4000  -3.9323  15.7822   3.0792  11.6653  12.6922
#>   -6.1660   9.9804 -14.1163   5.2564  -0.7400   1.7812   5.8941  18.7334
#>    1.1702  -4.9642  11.3427   1.4485  -4.4050  -2.4838  -4.9758  -2.9555
#>    0.6456  -2.2707   1.3142  -5.0772  -1.3803   2.3356   1.8466   8.7614
#>   -8.0263   1.3140  -8.3966   8.5826  -3.9063   2.6485  -6.4420 -10.2715
#>   -2.5822  -1.2052  -3.4047 -12.7640   7.1776  12.7286 -11.7339   6.1092
#>   -2.2696   1.8839  -5.2002   6.6378  -3.6410  -6.1426  -8.6698  12.1456
#>   -2.0754   7.3213 -13.0931  14.6557  -9.3843  -7.5568  -2.2852  -1.6569
#>   -2.7114   4.3902  -3.7837   3.2631   3.2648   2.0138  -3.9521  12.3818
#>   -2.9753   5.4404 -12.0957  -5.8171  -1.0358  -4.8458  -1.3635  16.6553
#>    7.6300  -9.2480   2.4145   0.0901   9.1824   1.3303 -12.9548  -4.2262
#>    4.2330  -5.3377  15.7440  -8.6310  14.9056  10.5762   6.5879 -13.3675
#>   -3.1262   1.3693  -3.9740  -4.6270  12.7288  -7.7031 -12.9530   8.2955
#>    6.2047  -4.6986  -5.3558   8.2403 -14.9823   2.0938   2.9226 -12.8182
#>   -6.2074  -1.7108   6.3271   5.8962  -0.0935  12.4702  10.6537  -4.3552
#>   -8.0372  -0.7439   1.1169   4.0019   2.7953  14.3633   1.0501  -4.7165
#>   -2.3026   1.3182  -4.4846  -1.9970  10.4912  -1.1024   5.7657 -14.9563
#>   -4.0054   5.0974  -2.0993   8.5337  -1.3030   6.0904 -17.1177   6.4852
#>    3.8095  -1.0041  -0.3285 -17.8837  11.8292  -6.1177 -10.1121   7.9989
#>   -4.4001  -1.8596  -0.0447   6.4914   5.0006   9.0896  20.4376  -4.7406
#>    3.1975   6.1727  -7.9561  14.7088  -4.5179  -7.6783  19.7912  11.7032
#>    1.5940  -8.8393   8.2182   2.3810  -1.0912   5.5675  -1.8522  -6.4732
#>   10.1626  -7.5913  12.8213 -11.9773  -5.3468  -3.2771   6.5014 -17.5678
#>   -4.3289  -1.8159   4.0073 -13.5022  -1.0243   8.8708 -15.6101   0.2883
#>    6.4651  -4.6910  -0.7458  -3.4712  -6.3350   7.2316   5.6005   3.8094
#>    7.8685   2.7562  -2.7832  -7.0178  -0.7484 -14.6482   1.3635 -11.0569
#>    0.2353  -5.1142  12.8901   2.5939   1.7955  -2.2733   1.6770  -7.5951
#>   -2.1378   2.7931   3.7595  -5.4788  -2.1671   9.3933   3.0939  12.6877
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]