Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.4530  -7.0836   3.3076   1.1375   0.1231 -13.8400   1.3578   5.5193
#>    1.6031   8.9584  -6.2451  -5.5751  -6.4610  17.9367  -6.3572  18.9505
#>   -4.0052   9.3710   7.2589   7.3985 -10.3405  -4.1738 -10.7755   8.0521
#>   -4.2186 -12.6743  13.1621   7.2011  -4.8641  -6.0464   3.1514 -14.0060
#>   -1.7729   4.3488 -20.3952   0.1922  -7.0000  -4.5963  -8.6457  -3.1107
#>    0.3919   4.0410   1.9976  15.5950  -5.1758  -6.6834  -5.3648   1.6212
#>   -3.5886  -8.6266  13.2429  -2.5172   5.7328   6.0606   1.1463   0.1591
#>    5.9022  -4.0287  -2.2271  -5.4668   5.7501   4.5156   3.3065  -0.5015
#>    3.0460  -5.0604   5.7196   8.4920   7.7120  -2.7529   9.3053 -10.9905
#>   -1.5421   9.6467  -5.5342  -0.9829  10.3792   1.4111 -12.5927   1.7768
#>   -2.6574  12.9667  -4.8921   2.4505 -14.8393  10.7714   1.7936  -5.1031
#>    2.7862  -3.5489  -0.6670  -1.2283  16.5830   0.9563  -1.3024 -12.6580
#>    1.7315  -1.2879   0.4022   8.2943  11.2533   0.1218  11.4180   4.1661
#>   -2.2058   0.9549  -8.7733 -13.4686  16.7526   5.9613   5.4653   1.4369
#>   -0.9376   2.3905  -5.5687   3.2664   0.3383  -5.7940  16.6673   7.1763
#>   -7.4313  10.7434   7.1466 -12.9850  -4.0843  -7.2933  -1.2282  12.5582
#>   -0.0367   7.6844   0.5717  -8.2872  -3.3831   9.4727  15.0716   8.9037
#>   -4.1580   1.9517   8.8590  -4.5962  -5.8809  15.3052   7.2329 -16.8429
#>   -1.2323  -7.0765  -4.3312   3.0198  -2.0258  -3.6481   6.7501 -12.3216
#>   -3.1879   8.7796   5.5590  -5.7169   1.2788  13.1534 -22.1695   7.7915
#>    2.1794  -8.3249  -4.7451  -8.0038  15.8316   7.6717   9.4537  -4.9073
#>    1.5179   5.3600 -13.3748  -3.3502  -3.2915   3.5912  -2.9567   4.9018
#>    1.8715  -0.3845  -3.0624  10.2753   4.8761  -9.6888   6.2086  14.2500
#>    5.6515  -1.5127  -8.9259  -1.4883   3.7650  -8.1382 -10.7556  -6.9819
#>    4.6996  -0.8743  -0.4773   1.4318 -11.1443 -15.1804 -28.4051   3.1165
#>   -3.5416   6.5673   4.6777  -0.5578  -7.8806  -4.9380   9.0952   6.3927
#>    0.8603   0.9950   2.6630 -21.7487  -1.9064  17.1416  -9.2358   0.0068
#>    2.6295  -0.6041   5.1409  -4.9189   0.1111   3.4250   5.2897   8.0524
#>    2.2129   3.6805  -6.9082  -4.8045   5.3268   0.1714  -4.9280   5.0507
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]