Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.2707   0.9681   6.2281   0.2862  -0.4722  -4.4274  -3.0652  -6.6904
#>   -6.0769  -2.0407 -12.1002  -3.2999  -8.0523  -0.0624 -12.7006  -1.8923
#>   -0.0926   1.6998   0.6983   2.5379  12.5433  -2.7473  -4.1838  -1.1362
#>    2.4377   9.8490 -11.1314   5.5877   1.9828   3.6974   5.3678  22.0381
#>    1.1005  -7.0384  -8.4385   0.1843  10.0808  -0.1104  14.7562  -4.5579
#>   -0.3692  12.9798  -4.7469   5.9548   0.9734   0.1931 -11.0076  -2.5149
#>    2.1385  -5.4307   9.3397   4.0006  -2.3028   0.5960  -2.9632   3.0266
#>    0.5554   2.9426   3.2110   0.1330   6.6958 -13.3339  -7.1171  23.1606
#>    0.5882   1.2302  -0.0464  -1.1022 -12.2347  -1.3336  -3.5146   5.0157
#>    8.7809   2.8458  -1.3377   1.3969 -16.8752   4.8565  -6.1647  11.6719
#>    3.5316  -8.0075 -10.3651   3.8378 -11.8735  -0.9748   2.0882  -0.8562
#>    0.1057  12.2215   6.7769   4.6780  13.4610  -0.5469   0.3865  -6.3521
#>    7.8683  -2.4959  -6.6410   9.6707  -9.6234  -2.4105   9.8731  -6.1836
#>    7.3145  -8.6021   6.0755  -5.6275  18.2386  -8.3264   0.5352   8.6626
#>   -0.7751  -2.4944  -3.6080   1.6004  25.6541  -1.3240 -10.7633   2.0739
#>    1.0166  12.3432   6.8951  11.7309  -9.5570  -5.1630  -4.0778 -10.0180
#>   -5.3651   1.8544  15.7798  -7.1146   7.3967  -4.8988  -5.6644 -12.9910
#>    0.4472  12.0767  -9.4038  -3.5651  -8.2332  -7.1683   2.0210  16.9302
#>    4.0938   4.8297   1.9865   8.5623 -15.5798 -14.4567  -0.9546   8.7137
#>   -7.1825  -3.4649  -1.9093  -1.1419   6.2738   1.5046  11.0985  15.9848
#>   -2.7659  -5.7495   9.2341  12.6658   3.3336  -2.2150   9.5206  12.1143
#>    8.6570 -11.0306   8.9952  -5.1497 -10.1499 -10.8043  -4.4627   7.1410
#>    0.2190   6.3411  -2.7302  15.8593  11.0464  -7.6910 -26.0335  -8.8873
#>    4.1782 -19.9224  -7.5030   4.0363  15.8982 -12.9656   2.4903   2.0078
#>   -0.4307   0.2850 -12.0327  -5.2158 -14.8687  14.9600  -0.8561  16.2560
#>    5.7115  -6.9812  -8.5297   0.8858 -22.8916  -0.8671   0.3890  -1.3942
#>   -5.9128 -13.2770  -8.8996  -3.2372  -3.2830   4.2207  -8.8491   8.7178
#>  -13.6338   0.1836  -5.6382  -4.2602  -0.9205  16.3538  18.2029 -10.4168
#>    4.3603  -1.4734   8.4530   8.4178   2.7410   3.7376  10.7226   1.0213
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]