Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.0099  -6.7324 -12.7169   6.1028 -11.2961 -10.8485   1.4566  -3.8158
#>    1.3901   2.1524 -13.1336  -7.3648   0.7777   5.6911   2.0829 -10.3212
#>   -3.9980   3.0336   4.4354   0.5670  -4.6699   0.2340  11.2585  -5.0219
#>    0.6206   1.7863   8.9816   4.1237 -11.9049  10.2049  -2.8475   8.7133
#>   -0.1007   6.3424 -10.9984   9.1876   4.7162 -18.8348  13.2146  -9.4839
#>   -0.0799  -5.4141   2.1122  10.6414   2.9648  -6.6992  12.5961   6.1616
#>    5.5017  -4.3546  -2.2505  -6.8775  12.1981  16.4375  -2.7288   0.3450
#>   -2.4058   0.4582   6.4347  -2.7089  -0.6626  16.9297   6.6064  -0.2065
#>    1.6978  -4.2845   5.2031   9.0229   8.3733 -10.2822  -2.1403   3.5471
#>   -0.7498  -3.6227   0.1164  -4.9098 -10.6262   3.4984  -3.9863  -8.8427
#>   -2.9389   1.0804   0.0144   0.5554 -11.9851  -0.3255 -19.5546   3.1760
#>   -2.9538   8.7740   2.8702  -5.0208  -1.6582  13.7480   0.5508   0.9759
#>   -4.2456  10.9440  -1.7275  -1.3651  10.3637  -3.5556   2.2502   3.6728
#>    2.1311  -2.2895   3.5985  -5.9496  10.2536  -4.9933  10.6267  -6.0691
#>    2.4605   7.6933   6.5470   9.9547  12.7939  -6.2139   0.7852   2.9621
#>   -0.3692   5.7254   3.0155  -4.7373  -1.4345   2.4041 -14.4237   1.1280
#>   -0.6462  -7.2485 -10.0071  11.7761   3.1531  -2.4382   1.4428  -8.6639
#>   -0.4645   0.0439  -5.4247   1.4887   1.0237  -7.0003  20.7605   1.4609
#>   -2.3717  -1.8996   5.0806  -1.7044  10.3009  12.0018  -3.7649  10.2501
#>   -4.6883  -6.7207  -9.1565  14.1027 -11.5027  -2.2151   0.8941  10.0027
#>    4.7149  -0.7449  13.5267   3.4094  -3.1599   9.1213  -5.5953   6.9705
#>   -3.5990   5.6758  -1.1481  -6.3205   1.9060   9.0628 -13.6891  -6.0128
#>    2.0179   2.4978  -0.1741  -6.7093  -6.5602  -6.4065 -16.8715  -2.4228
#>   -1.2489   4.6528  -4.4241   0.6286   0.6987  -5.1665  -3.4116   1.0132
#>   -6.0362  -3.4685   4.6282  -3.4481  -1.0279  -6.7894  10.0255 -22.0202
#>   -0.0459   1.0887  -0.1052  -1.4305   0.4974 -10.9161  11.6561  -1.2734
#>   -2.8733  10.4142   4.1242 -14.1048   8.4167  -3.3876   8.1035  -8.2704
#>   -7.0780  11.2812  -0.0534 -10.9579  -3.4537  -0.8917   2.0804 -11.9659
#>   -3.2724   5.3138   1.4814   5.6013   7.6683   3.3176   6.6691  -2.5940
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]