Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.1119  -1.3492   8.7028  -0.1433   5.2230   0.9347  22.3440   2.6281
#>   -1.1523  -4.2780  -6.2482   6.6165   7.4512   5.5765   2.3996   7.5356
#>    0.4154  -5.1160  -1.0623  -1.9408   3.9859  12.7834   0.4419 -19.8753
#>   -6.6744   1.1235   9.6431  13.0888  13.4011  -7.0812   8.1935  -3.5122
#>   -2.1360   0.6304  -2.0379 -10.3324   9.0268   8.0968  -9.1230   3.7009
#>   -1.7377   4.9753  14.9810   9.4326  -3.9426  25.5676 -12.6450  -1.6837
#>   -7.0948  -1.3337  -0.0719   0.2959   0.1573  -4.8103  -4.0817  -9.5386
#>    2.5313   4.6229   9.7293   4.5796  -1.6266   4.3516  21.8566 -17.5260
#>   -0.8993   7.8335   5.1093 -19.5103  -9.4632  -5.8785  -7.0321   1.8446
#>    8.6140  -6.7114  -2.8087  -4.5322  -7.2499  -1.7193  -2.9788  -5.0971
#>    3.7142   1.1426 -12.0045   4.2593 -12.4169   6.4605  -7.7673 -11.2639
#>    0.2695  -3.3257  12.2049  -3.3675  26.4677  10.4265  -6.9816   3.1913
#>    4.1610   0.9725   3.3115   1.3512  -8.5169  -2.8756  -0.4195  -1.7298
#>    1.5312   5.6999  -2.8029   5.6290   4.0897  -1.3276  -9.0432  -7.5716
#>    3.4637   9.4145   4.4941  -1.3601 -12.1559   0.9973  -1.4199 -12.7275
#>    2.8249   0.4919  -2.7011  14.3567  -1.5622   1.2046   3.6733  -5.5297
#>    2.8528  -4.4213  -4.6481 -14.4708   4.8536 -24.6023   4.0641   4.3342
#>    2.2847  -4.8493  -7.0549   5.4558   1.2789  -8.5863   9.3584  -7.1309
#>   -0.6523  -0.8845  -2.9695   9.1732  -4.3703  -8.4131  24.6185  -7.3608
#>    3.8019   4.2019  -9.4901  -5.0268  -3.1375   2.9444   6.4196  -2.8716
#>   -3.9731   3.7307 -10.7628   6.1127  -0.6252   3.9651  -5.3724  -2.3534
#>   -0.4376   2.4239   3.0689  -2.9281  19.1978   9.4092  -7.2327   1.7383
#>    3.0103   1.8175   4.4014  -6.8356  -6.1958  -0.8444  22.9520 -13.0858
#>    2.8703   8.9274  -8.4476   1.5888  -3.0020   1.8215   1.4986   8.8502
#>   -0.6512  -9.1165  -0.2314  -4.2566   4.6336   2.2700  -7.4648  -9.3272
#>    0.6663  -8.3651   5.2198 -10.8938   1.8013 -10.4940  -3.4126   5.1234
#>    2.8964   0.1519  -0.1908  -0.1348   7.1240   2.1548  -5.4786  -0.5675
#>    5.6951   3.4154  10.6410  -1.8537  -6.5816  -0.6078  10.3713   1.6402
#>    3.9096   4.8081   4.9571   8.6125   6.9468  -7.2274 -13.9946  12.1558
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]