Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.7633  -0.4066  -5.4262   1.0119  10.3245   3.9090   9.4194   8.5052
#>   -3.4906  -4.3156  -3.6824  -0.1170  16.4432  -0.5658   3.9214  -4.8505
#>   -4.1516  -2.8036  -2.6965  -6.4818   5.8247   7.2452  12.7185  -7.0505
#>   -0.2995   2.7455  -6.2624  12.9158  -0.6493  12.6060  13.4230   9.0471
#>   -3.8213  -1.1062   4.3151   6.4032  -4.8163   5.8566  -7.2581  -3.5354
#>   -1.5176   7.8978  -8.2270  -5.8210  -1.3576  14.9066  -2.0314   0.1213
#>    0.3201   0.2676  -3.0465   9.1620   4.8540   3.6864   6.7487  14.0777
#>    2.7083   2.1232   8.6399  -4.3806   1.6992  11.0694  -8.3944 -15.2192
#>   -1.7217  -5.7346 -11.3640  -1.0837  -2.0468   7.7126  22.1238 -10.6028
#>    3.4561   9.4812   0.6970   7.6235  -7.5735   3.8392  -9.5088   2.3685
#>   -6.2431  -7.3527   1.4494   5.1587  20.6117   0.4495   5.4944   2.4312
#>   -1.6148   0.3689  -1.4960  -5.8839   9.9896  -4.0661   5.0336   5.0668
#>   -1.3536  -4.0555   5.2525   0.7618   6.0916   0.1353  -2.6963 -18.5372
#>   -2.2748   0.2184 -12.0871  -4.5457  -8.4278   4.7251   4.4607  21.8063
#>    0.2794  -4.2141  -2.8534 -22.4807  11.1820  -0.0348  10.7024   5.5726
#>    6.9287  -6.7204   9.8689  -2.5349  -6.6221   3.4983   2.6455  -4.9244
#>    0.6804  16.2317  -4.0965  -3.0850   1.6728 -14.0464 -13.3728   6.2868
#>   -1.8283  -0.9020   5.2774   1.4420  -7.1427  -2.3152   7.2119 -11.1360
#>    3.8683   1.0978   5.0844   0.8375   6.4915  -2.4449   3.6429  12.2712
#>   -3.8096   0.1533  -0.7555  10.1524 -19.6714   6.2519  -4.2116  -0.6256
#>   -5.7951   6.9021  -0.6680   0.2205  13.3611   5.9409  -6.6866  11.6258
#>    1.3099  -5.3944  -0.6378  -2.3678  -7.5177  -0.4137 -13.4284 -10.3827
#>    4.9188 -14.5958  19.5321 -11.6873   9.9612  -1.0896   6.7839 -11.1537
#>    1.4134   5.1471 -12.3998   7.8286   4.3600  -0.3835   2.2105  21.9740
#>    5.2478   6.1390   0.2086   5.1614 -11.5267 -10.8983   8.9066  -0.9092
#>   -0.5734   6.4011  -0.9907   1.9613  -0.7724  18.6860  14.2338 -10.9760
#>    3.7870   4.3166   1.4887  -1.2169 -17.4936   7.4023   3.1522   7.6314
#>   -0.1413  -3.0996   7.2322  -4.5234  10.7147  11.7332  14.3828   1.9326
#>    3.3685  -2.8909  -2.2224   0.1273   7.9533   1.2029   1.1790 -11.8652
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]