Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  2.7100e+00  6.7977e+00  1.2342e+01 -9.0474e+00 -1.9423e+01 -1.8584e+01
#>   4.3962e+00  8.9851e+00  4.0194e+00 -2.7276e+00  4.6491e+00  2.8784e+01
#>   4.5053e+00  2.7480e+00 -4.3642e-02 -5.5070e+00  1.1639e+01 -9.0109e+00
#>  -3.0660e+00 -2.1444e+00 -1.1666e+01 -1.5466e+01  2.7274e+00 -1.3927e+01
#>  -5.5063e-01 -6.3721e+00  1.1334e-01 -8.0707e+00  2.7879e+00  5.2224e+00
#>   2.4642e+00  1.0483e+01  1.2795e+01  8.3436e+00  1.9515e+01  6.0101e-01
#>  -1.9767e+00 -7.3966e+00 -1.9117e+00  5.0268e+00  4.1281e+00  1.5231e+01
#>   5.2492e+00 -7.3062e+00  7.3602e+00  1.1827e+00  1.3879e+00 -3.8689e-01
#>  -4.2321e+00  5.9662e+00 -3.3582e+00  5.7489e+00 -3.4906e+00  1.1395e+01
#>   4.5789e+00 -3.8171e-01  1.2519e+01 -2.1103e+00 -6.3840e+00 -4.1277e+00
#>  -2.4272e+00  5.3165e+00  2.3651e+00  5.3399e+00  2.1278e+00  1.0658e+00
#>  -4.7012e+00 -5.8759e-01  4.8259e+00 -5.6969e+00  8.8392e-01 -6.2990e+00
#>   5.3253e+00  3.9875e+00  1.9085e+00 -2.4508e+01  1.4061e+01 -1.8483e+01
#>   6.1309e-01 -2.3213e+00 -3.4901e+00  1.6531e+00 -6.3424e+00  1.9233e+00
#>  -5.7581e+00 -3.4444e+00 -6.4757e+00 -2.1694e+00  1.2256e+00 -6.7122e-01
#>   3.6217e+00  2.5022e+00  1.6152e+01  3.6723e+01  2.0496e+01  2.7724e+01
#>  -3.9997e+00 -2.7301e+00  2.9967e+00  1.3444e+00 -2.3035e-01 -1.3293e+01
#>   3.8777e+00  4.7860e+00  7.2840e+00  2.2925e+01  8.9926e+00  4.8426e+00
#>   9.1388e-01  6.2928e+00 -7.2341e+00 -6.5472e+00 -6.1216e+00 -3.0901e+00
#>  -2.4872e+00 -2.6115e+00 -5.0797e+00 -8.9126e+00 -1.1058e+01  3.9579e+00
#>   1.2263e+00 -3.0065e+00 -3.7728e+00  1.1128e+00  1.7121e+00  1.3956e-01
#>  -2.3187e+00  7.5939e+00  1.2621e+00 -5.9513e+00  1.3740e+00 -1.1608e+01
#>   3.9674e-01  1.0023e+00  3.5695e+00 -7.2344e+00 -9.4787e+00 -1.9710e+01
#>   2.1297e+00 -2.4544e+00 -1.1175e+01 -4.9525e+00  4.8176e-01  1.3675e+00
#>   6.0443e-02  8.2069e+00 -1.4115e-01  5.1271e+00 -4.2783e-01 -1.6032e+01
#>   3.3382e+00 -1.4301e+00 -9.9280e+00 -3.4697e+00 -2.0050e+00 -1.3285e+01
#>  -5.3349e+00  9.6996e-01 -1.8291e+01  6.0579e+00 -1.2619e+01 -9.6090e+00
#>  -2.3409e+00 -1.0598e+01 -1.3018e+01 -3.0907e+00 -2.5787e+00 -1.7688e+00
#>  -3.8229e+00  8.3668e-01  1.3913e-01 -1.6715e+00  1.3219e+01 -2.3629e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]