Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -7.6229  -0.0376  -3.8575   0.9432  -3.2636 -16.8776   8.0091   4.3315
#>   -6.0199  -1.2554  -6.1098  -6.4685  -3.0709   7.0731  11.4618  -4.4127
#>   -6.0679  -1.9764   7.1476  12.6744 -12.0190   7.4959 -11.9752  -6.6647
#>    4.9711   0.6631   7.3544  12.9725  -3.3256 -12.9630 -11.3023  12.4263
#>   12.8659  -8.0618   0.5775   1.1439  11.6230   1.1447  -9.2313   5.7501
#>    2.2880   7.9900   9.2989  -6.9665  -1.2257  11.8680   1.7873 -19.7614
#>    2.6910  -4.9148  -3.9234   3.6141   1.2371   3.6021 -23.0792   8.3807
#>    5.6485   4.0951  -7.8176   0.4238  12.1291  11.1190  10.3251  11.3286
#>    1.9282   6.9376  10.7227  -6.8051 -11.1494  -4.2264 -18.1214  -4.1325
#>    2.3460   6.6535   5.4583   4.4013   6.1590  -2.4157  -0.0429  -1.1077
#>    2.2105   1.7309   1.5117   1.3531  -5.7761  -4.4722  19.2400   0.8792
#>    0.2868  -1.2976   8.1621 -13.6007  -0.1472 -15.6090  16.8397   3.8543
#>    1.6614  -7.6626  -5.2064   3.3123  -7.4541   0.2244   0.0871  13.4793
#>    0.0305  -1.6073  19.7100  -0.7644   8.2348  -4.4115 -10.6037   3.0903
#>    1.3819  -0.7616  -7.1216   1.1329  -1.9631  -1.8288  14.2685  -8.1606
#>    7.3681  -2.2075  15.1161  -0.2251 -10.0796  18.3353  -6.6872   0.2233
#>   -4.7840   5.2799  -4.8070   4.8832  -0.8456   2.2113  -4.8509  -0.2980
#>   -0.5072  -4.4734   4.9365  -3.0391  10.6807  -3.8563 -14.2630  -4.4678
#>    5.7535   4.3309   2.9138  13.6993  -7.0815  11.9785  -5.2975   6.5786
#>   -7.2137   3.3801   5.9135  -4.1112   7.3289 -12.6370  -4.1445   5.7932
#>   -4.5602  -6.9140  11.9746  -8.0104  -5.4034  -1.9183 -12.6072   7.0647
#>    9.2398  -3.0755   4.3594   1.2511 -18.5151   8.5591 -15.4981  23.8023
#>   -0.0369   3.1493  -1.5217   4.4356  -6.0388   5.9520  -4.6505  -5.8473
#>   -4.0147  -5.5976 -10.7512   2.1679  -0.1437 -14.5024  -8.4795  -9.2768
#>    6.8979  -3.6417   7.5738   7.6178 -16.4395   8.0115  -4.5658  -8.8371
#>    0.5774   4.4006   6.9646  -3.3404   0.9004  18.9631  -5.2421  -5.8953
#>    4.1667   6.1264   1.7237 -13.5735   1.9874  13.1595  18.0587   3.5979
#>    4.8677   8.0231  -8.9116  -2.1086   1.6757   6.1703  -3.9868   2.7408
#>    5.1632   2.8005   1.9086   0.3746 -10.0527   2.2161  10.6442  20.4043
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]