Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.8283  -1.4634   4.7686  -0.4974   3.6659  -0.3557   9.3112  -9.8771
#>   -4.4738   3.3195   1.1378   6.2649  -1.6978  -5.0502   3.0802   5.2151
#>   -0.9098  -3.8492  -6.1236  -3.4219  -5.0766 -15.6191  -2.2014  15.0180
#>   -0.6789 -10.3511 -12.5977  -4.5097  15.8431   0.8718   7.4315  11.9772
#>   -5.0025   7.7004 -10.8255  -3.3367   1.8074  -6.8129  -7.8853  -8.2960
#>    3.7679  -7.3193  -5.6582  -5.4387   7.1081  -7.3071  -9.8584  -2.5431
#>   -3.5985   6.4602  -3.9483  -7.0935  -7.4436  -3.3628  22.7420  -0.6906
#>   -8.3051  -5.9475  -7.9970  -1.7950  -5.7763   2.3342 -13.7356   7.7228
#>    3.7846  -7.2027  -1.9996  -4.0366  -5.9217 -13.5321 -11.8741  17.4684
#>    0.1275   0.0474  -0.6587   3.2954  -0.0823   2.4921   3.5741  10.2507
#>   -3.1172  -1.2868   1.7571   7.1152  -1.4348  -2.4732  10.9952  -0.8950
#>   -3.6571   2.3037   7.9194  -7.5086   6.4852  16.0418  14.6227   8.6402
#>   -4.4940   1.0208  13.6047  -1.9370  -9.0564   6.4504  -9.7101   4.0175
#>    8.9360   1.6402  -1.9555 -17.3444   5.9116   2.1937 -11.9738 -14.5644
#>   -1.2256 -10.3034  -9.8928  -0.7898  -1.1181  -0.7292   0.6146  11.7041
#>   -2.9610   1.4348   1.0208  -3.5277   6.2295  14.8298  -3.2984   1.7376
#>   -1.2824   2.4260  -0.0134 -13.8197  -6.9327  -6.0355 -11.0048 -14.5450
#>   -4.9117   0.6051  -7.6916   9.3724   6.2505  -0.1082  -5.4665   6.9191
#>    4.2614   1.4604   7.6844   4.3555   2.5527  10.8739  13.3989   3.8637
#>   -0.5413   6.3720  -7.1118   0.6651   5.5830   3.3112  -7.9687 -16.5121
#>   -0.2506   5.3135  -3.3427  -1.5802  20.8613   1.9272 -10.0070 -13.9053
#>    2.0105   3.0865  13.7840  -3.7128  15.4719  12.8612   6.1347   1.1096
#>   -2.4462  -0.7404   2.3681   1.6852  -8.2217  -5.4826 -10.8973  -9.5639
#>    1.8280  -5.6593   1.2026  11.5653   0.9697  12.1972  -3.2926  -1.7525
#>   -2.4086   2.6756   7.7816  -2.6347  -1.0449   2.7095   9.1759 -12.3064
#>   -3.4295   7.2011  -1.0442  10.7893  -1.1440   0.6235  11.7251  15.2595
#>   -5.4330  -0.6345   8.8065   6.1358   0.7662   2.6784  11.0634  -0.0146
#>   -3.1996  -2.1931  -2.4949 -12.2419  -9.1966  -7.2318 -11.1886   4.4544
#>   -1.6836   5.9058  -3.0673  -4.2478   9.7889   1.9508  -5.5516   8.5107
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]