Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.3747  15.9033   0.4768  -2.6755   6.7284  -7.3816 -15.6958   6.8242
#>   10.6894   3.5815 -10.3444   9.0293  -6.9775 -22.7653   1.4509  31.1745
#>   -3.5512  -8.6297   0.6642   2.8404  -1.2514  -7.7232  -4.3754   8.7740
#>    1.9240   6.2739   4.0834   2.9230   0.5887  -4.2586  -0.1142   8.2952
#>   -1.9130   9.5068  -0.2349  -8.0046  -3.7827 -12.1306  -0.3777   1.7831
#>    9.0320   1.9504   3.1722  -4.4648  -9.2607  14.7951   7.8378   0.8108
#>   -9.1167   3.1305  -5.9777 -12.2474   9.4072   5.9302   2.2284  -9.7980
#>   -4.0332   1.5046  -8.8957   8.5845  -9.8176  -7.0317   0.3065  25.9427
#>   -2.6412   7.2903  -5.6163  -8.9355  -1.9559  12.1492  13.9454 -13.1999
#>    9.5163  -5.6252  -1.5995  -0.1171  17.3303   4.8911   5.9131 -11.4430
#>   12.2939   8.4839   5.2033 -13.2615  -6.8339  -5.4383 -10.6961   8.2645
#>   -1.2060  -8.3278  13.6627   0.9225  -8.1519  -6.7554  10.0103 -15.4426
#>    2.5854   9.4131  -6.6516   0.5342  -8.8471  -2.9042  -3.8749   9.1625
#>   -4.2958   5.7866  -8.6474   1.6595  20.0803   8.7255 -13.0163   9.9673
#>   -0.7231  -4.2463   2.7273  13.8689  -8.6730 -29.2925  10.3195   5.6369
#>    0.6531 -16.4500   2.8829  -2.0957   2.5289   5.2439  10.7335  -2.1677
#>   -2.4488   0.7228   6.7960  -4.4589 -11.7617   2.7952   4.4541  -1.8219
#>    2.1922  -2.5827  -4.5427   2.5442  12.6033   7.6940 -12.3160  -9.6387
#>    5.7127  -4.0012   1.9856   4.4179  -7.2168   0.5156   7.6099  -5.2401
#>    5.5481   5.6757   0.2280  -0.4747 -11.6897  11.3243   5.6797  -8.1760
#>    3.3365  -3.8053  -5.7144   2.3351  -6.2424  22.7791   0.8072  -4.5220
#>    1.0368   9.2816  -0.9899  -3.3127  -7.1763   0.8943  10.1507  -4.4346
#>    0.2358  -3.3585   8.3248   0.7058  -7.5823  -4.3348   4.4155  -5.3751
#>    0.3692   9.3537 -10.6486  -8.5941   4.1804  -6.1700  -8.7839   1.9123
#>    5.6526  -0.8524  -0.9351  15.1014   3.8499 -12.9687   4.5068  15.8607
#>    1.7125  -7.8110  11.3738   9.8831  -4.5059   3.2644   9.6453  -2.2368
#>   -4.4326 -11.2479   7.5599  14.4029  -7.1123 -11.3369   2.1015  -0.0923
#>    4.6190  -9.2666  -6.6991  -2.3939   2.2583 -23.7156   9.1093  -7.4530
#>    5.6177 -14.7427  -3.2459   3.8941   6.5177 -19.6192   1.3059  11.6781
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]