Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.8103   3.7872   3.1298  -2.4401   3.0832  13.8301 -17.2690   9.8202
#>    1.8926   3.4611  -7.7970  -8.2572  -1.8939   1.9911 -16.2066  -4.8687
#>    1.8939  -0.5323   7.2677   7.6693   9.3160  12.5863 -11.2722   8.5978
#>   -6.1721  -5.0883  -1.7033   8.6094 -25.9116   0.1497   9.1140   2.6856
#>   -2.2663  -1.5544  15.4378  13.9446  11.6764 -15.4888  17.9778  -0.4563
#>   -3.6060   3.8340   9.7362   4.0964  11.4814  12.1601  -8.8619   8.2265
#>   -0.8885  12.0326   3.4998   9.2465   2.8313  -4.9707   1.4666   6.8501
#>   -3.1565   1.1245  -1.9098  -7.3397  -6.8876   8.6886  -0.3726  -3.6938
#>    0.6396   6.1073  -2.7629   0.6423 -11.8193   2.9703  -7.9732   6.4288
#>   -4.1444   1.3250  -5.1928  12.9373 -14.7642   3.9226   2.0145   3.2857
#>   -2.1768   4.3243  -9.7508  18.5673  -3.5619   3.5238 -11.0171   6.2927
#>    2.3577   3.3574   3.4396  -4.9542   0.9649   6.5226  -6.0263  12.4510
#>    3.5267  -4.0851 -12.1776  -2.8879 -11.2938  -4.8121  -8.5535 -10.8793
#>    2.8137   6.1129   0.8516  -1.4462   4.8136  -2.9512  -1.6529  -1.5404
#>   -0.1082   0.5565  10.8752  -8.3569   6.5210  -8.5631  13.6279   6.0327
#>  -10.2988  -6.9755  -8.7077  -5.8597  -8.2611   1.9173   3.6351   1.0198
#>   -4.5993  10.4738   3.3605  13.0443  -5.0394  11.2646 -10.0991  11.1553
#>    7.7507   7.6012   9.7352   4.9345   4.9955  19.6275   3.2989  -1.8363
#>    0.7650  -1.5593  -1.0492  18.1841  -1.3482  -0.3738   0.0825 -12.4137
#>    4.8513   7.6185  -0.1124  -2.0091   8.0370   2.7701  12.9951 -12.6023
#>   -3.6116  -2.9414 -17.8317  -2.7919   6.1753  17.8017 -17.6308   8.9569
#>    1.8508  -2.1738   4.9588   1.3750  11.6430  -4.9894   6.7384  -4.5370
#>   -1.2801  -3.8551   3.8709   1.4100   1.5435   4.2413  -6.7108  -5.0862
#>   -1.9310   6.4743   6.1569   6.8288  -2.1353 -11.3146  10.0271  -8.3089
#>   -1.5216  -2.5112  12.4731  -1.1005   8.5411  -6.4598  -2.6251  -7.7402
#>    3.7927  -1.9690  -1.2740   7.0931  -1.5949  -5.2803  14.2837  -6.8095
#>    0.7660   3.4584   5.3409  -4.0027  -1.7087  -4.4103 -11.2704  -5.1960
#>    2.2907   2.9272  12.9356  -3.9878   3.4541   4.0408  -4.1233  -2.8870
#>    0.5257  -0.4787  -0.5716 -10.6505   6.2252 -13.6758  -8.6266   4.6844
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]