Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.2992   9.0526  13.1275  -2.8508 -29.8450   4.0811  -0.5589   8.7344
#>    3.1098  -3.2784   0.8681 -13.2914  -2.5631  -1.2237   9.8639  -1.8823
#>    1.0371  -8.9765   3.2584   2.3142  -5.1093  -2.8243   2.0300  -2.0924
#>   -7.5857   5.5674  13.6972  14.1949  -9.7491 -15.8117  -8.9161 -13.8598
#>    1.3391 -15.1660   5.0185  -3.0976   8.0432 -12.4719  -7.2395   2.9760
#>   -1.3434   2.2198  -0.0879  -2.3606  -2.2911  -2.2249   0.6255   1.7841
#>   -1.2029   3.0552  -0.8647  -3.8435   5.1248  -8.2567  -6.2036  -9.6716
#>    4.6973  -5.6154  -0.3189 -11.1135 -11.2510  -4.1856  10.8785   7.1831
#>   -4.6122   8.1054 -24.3057  -3.5188 -10.2197   4.1989 -23.0482   9.3880
#>   -6.5363   8.3766  -5.5209  -3.3545 -16.2447  -4.4594  -1.7863  12.5695
#>   -9.8592  -4.3572  -6.8892  12.3615   1.9316  -1.5192  -5.1566  15.4629
#>   -1.1806  -7.9518  -6.5524  -3.8923  13.7821  -1.6128  10.0795  -4.8039
#>   -2.2193   0.7055   2.4734   9.8941  -8.9675  -1.5527  -1.9520   5.2196
#>   -7.6626   1.9093  -5.7827 -13.4064   9.4132  11.1114   5.2993 -10.6703
#>    5.2318   7.9039  -2.8670  -2.9641   3.2852  -1.6008  -5.2175   2.5643
#>    2.2245  -0.7171   7.0327   9.0109  -0.0701   4.6273  -1.2953  -1.4634
#>   -1.0295  -4.2372 -10.2706  -3.8486   5.7971   4.5962  -6.1009  10.4042
#>   -2.2978   5.3294  -0.2275  -9.8479  -7.5469 -12.2268  -6.4225   8.7360
#>   -0.0316  11.7910   0.2372  -0.0475 -16.2580  -7.1267  -5.0565   5.5799
#>   -1.3730   1.1529   5.5130   4.1155  -2.2084  -3.6786  -9.4294  -4.6215
#>   -0.4551  -9.2690  -3.2865  -8.3246  -5.3493  -1.0762  13.5819  -5.0604
#>   -2.2345   8.6720   5.0315   8.0865 -15.7897   8.8763  -3.3448   7.3400
#>   -1.3604  -0.9662   1.1825  -0.0632  -7.4611 -15.8317  -9.4664   2.4308
#>    5.1649  -1.8699  -4.3629   7.4831   2.6499  -8.9496  -2.5214  10.5859
#>   -5.9233  -0.7043  -0.7938   4.9301   0.2490  -5.8611 -21.6334   4.5487
#>   -4.1066   6.3773   2.9997   5.4660 -20.5753 -14.3034 -17.5180   9.1582
#>   -7.0255  -1.1292   5.8823   5.2114  -3.5129  -3.7333  10.2341  -7.3523
#>   -1.1037   2.7877  -3.0403  16.3458  -8.0515  17.6022 -11.3326   7.9450
#>    1.7911 -12.7248   8.6123   2.6187  -4.4655 -13.3754  17.2879  -2.1635
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]