Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.0935   6.1847   6.4218   4.0726  -7.0648  -1.2492 -14.2253  10.2144
#>   -0.8127   4.4845   7.6183   2.4982  -9.7623  -1.0076 -15.0254  -0.2789
#>    0.5781  -4.4588  -1.3254   3.8620   9.4725  11.8435  12.3786   7.0595
#>   -5.4977  -6.3376   5.4770   2.4845  -4.8765 -14.6552   1.3288   3.8160
#>    7.8374  -4.4580 -16.0430   0.8820   9.2693   4.6303   5.5827  -5.2294
#>   -0.0845   3.5116   0.3542  -0.6420   0.0915  -9.9375  -0.4985   4.8783
#>   -0.6722   7.2351   2.0645   2.7994  15.9543  13.4722  -7.6077  -0.4564
#>   -0.4861   1.7881  -7.7582   7.7410  -8.0992   0.8724  10.7743   2.0404
#>    1.4589   0.8398  -4.0823  -7.7797  -0.2804  -4.8815 -15.2382  -9.0929
#>   -3.6871  -0.2876   0.4142   0.4951 -11.7859  -0.7282  -5.0354  16.1917
#>   -1.0351  -4.0745  -1.8011   3.4323  -4.2975 -17.9869  -2.8889 -13.0048
#>   -1.5431  -3.4921   1.0946  14.3591   1.2242  -8.9657 -22.0309 -10.4539
#>   -1.0538   1.7400   5.1948  -5.8870 -15.3565 -13.4283 -10.3659  18.4029
#>   -3.6261  -3.6268   2.5806  -7.6041  10.6751   5.8467  -3.4546   5.5034
#>   -4.0242  -0.9051   3.9008  -4.1617  -2.7873  -6.0977   0.4776   1.0319
#>   -4.9527  -2.6425   0.4613   0.3826 -13.7141  -1.4922  -8.2295 -22.8319
#>    0.7375  -2.5477  -0.5558   4.2575  -7.5197   4.2607  -3.4090  -6.7490
#>    1.2070  -1.4515  -5.8119  -2.4239  -4.1859   9.3119  -0.4197   4.5587
#>   -1.4977   1.3682  -1.4392   3.0992   4.5363  -2.3138   1.0938 -11.6422
#>    8.0130   8.4092   0.4727   2.6992  10.7222  11.0151   0.7806  12.0364
#>    2.7308  -2.3083  -0.1054   8.6210  -9.8345  -2.4582  -1.2105   8.8739
#>   -1.2090  -1.0054  -5.6896  -2.9276  -8.6470   8.1110  16.8977  -3.6230
#>    1.1892  -0.0519  -5.1269  -3.5814   6.7859   5.3875   6.1588 -10.8298
#>   -1.1360  -2.4732  -4.5223   1.0112   8.3970  -4.4377  -5.4956  -4.2757
#>   -2.9298   1.4128  -5.7656 -10.4074  -8.6156   7.4935  -3.9070   5.1971
#>   -3.6633   0.5480  -0.3021  -0.4003  -8.8662  -5.3786   6.5288   4.0416
#>    1.8379   4.3926  10.1437   6.8161  -1.8377  -5.2640  21.7917   1.8625
#>    3.6523   8.6698  11.4101  -1.1368 -15.3150  11.0509   0.7716   4.3105
#>   -1.1091   2.0318  -1.6320  -2.7847  -3.5472   5.7529  -6.8526  -4.1758
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]