Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  1.4234e+00 -1.5031e+00 -6.3616e-01 -1.6005e+01 -1.7653e+00 -1.1944e+01
#>  -2.6130e+00  2.1109e+00  5.0137e+00 -8.3263e+00  3.1316e+00 -6.7171e+00
#>  -7.7618e-02 -3.6638e+00  3.7729e+00 -6.2228e+00  1.4163e+01 -1.2355e+01
#>   8.0153e+00  1.8125e+00 -5.2152e+00 -6.1568e+00 -5.3453e+00 -3.1439e-01
#>  -4.8352e+00  7.3619e-01  4.5109e+00  1.3390e+00 -1.2117e+00  6.1506e+00
#>  -2.3411e+00  4.9159e+00 -4.8826e+00 -2.8050e+00  1.4312e+01  3.6273e+00
#>  -4.3436e+00 -2.9203e+00  3.0042e+00 -1.0826e+01 -3.2666e+00  3.9990e+00
#>  -1.7722e+00 -2.3305e+00  5.1786e+00 -9.0884e+00 -5.4102e+00  2.9818e+00
#>   1.2349e+00  4.8561e+00 -1.6917e+01  1.1370e+01 -1.1253e+01 -1.0506e+00
#>  -5.0320e-01 -5.4563e+00  2.2738e+00 -6.8237e+00  5.2816e+00 -2.3220e+01
#>  -5.2823e+00 -1.6666e+00 -5.6597e+00 -2.6170e+00 -6.4845e+00 -4.1117e+00
#>   1.0925e+00 -1.3915e+00  3.9011e+00  1.1433e+01  1.4513e+00  1.2619e+01
#>  -1.6974e+00  6.7357e+00 -6.2011e+00 -9.4999e-01 -3.8088e+00 -3.8713e+00
#>   3.2832e+00 -7.8925e+00  3.5728e+00 -3.6817e+00  3.9862e-01  1.4450e+00
#>   4.0368e+00  1.3909e+01 -4.8242e-01  7.8844e+00 -5.5844e+00  2.1601e+01
#>   3.8193e+00 -1.6119e+00  7.8622e+00 -4.2775e+00 -4.7867e+00  3.9014e+00
#>  -1.3596e+00  3.1102e-02  7.2094e+00 -4.2222e+00  9.9116e+00  4.0948e+00
#>   5.1430e+00 -7.9472e-01 -1.7387e+00 -5.0144e-01  5.8753e+00  3.1581e+00
#>  -4.0289e+00  1.2223e+01 -1.2749e+01  3.5464e+00 -1.4313e+00 -3.6972e+00
#>   7.3912e+00  6.3914e-01  5.8189e+00 -5.0191e+00  1.4044e+01 -1.5735e+00
#>  -4.0717e+00 -6.0249e+00 -4.7371e+00  2.7147e-01 -1.1632e+01 -8.3076e+00
#>   8.9647e-01 -1.8814e+00  5.1120e+00  1.4819e+00  9.9320e-01 -8.1805e-01
#>   3.3988e-01 -4.4315e+00  8.0483e+00 -8.9666e+00  3.7973e+00  5.9623e+00
#>  -1.1679e+00 -3.1561e+00 -2.7813e+00  1.5826e+00 -2.9887e+00  1.7741e+00
#>   2.6976e+00 -3.1211e-01 -1.0025e+01  1.8118e+00 -5.4081e+00  6.9539e+00
#>  -4.7193e+00 -2.1988e+00 -1.4907e+01  1.5257e+00 -4.3648e+00  1.0104e+01
#>   8.9952e+00  6.7141e+00  7.4326e+00  8.3896e+00  1.1992e+01  1.3093e+00
#>  -4.8924e+00 -1.9185e+00 -4.1971e+00 -7.3296e+00  4.8186e+00 -1.1817e+01
#>  -1.8650e+00 -5.4170e-01  3.7811e+00 -3.6127e+00  9.1621e+00  6.1826e-02
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]