Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.6362  -7.2872   1.5473  -2.0647   6.0405 -11.6195   6.6154   4.7842
#>   -1.8549   9.1026   4.7563   6.7166  -8.2138   9.7842   1.3200 -10.9113
#>    0.6259  -4.5922  -5.3729  -6.3752  10.2006  -8.6503  -2.4322  -2.5677
#>    3.5707   8.7554   1.5951   2.3449  -1.1080  -4.3936  -7.8857  -6.6179
#>    2.9494   0.0275   1.3784   8.2862  10.2633  -8.7769   5.6397  -4.2372
#>    0.7307   2.5421   0.2557  -0.7068 -19.3129   9.5518  -8.4017   2.5199
#>    0.3125   5.9163  -3.5965  -2.1321  -4.4314  13.6516   2.0914   1.2457
#>   -0.0129   0.4423   2.4886   2.7248 -20.6380   6.8210  -4.3477   7.0344
#>    4.2220  -3.8322  -2.0431  -1.3348   5.2730 -16.7331   5.0815  -1.1388
#>   -1.6947  -6.1532  -2.1219  -3.0582 -11.2658  -5.0755 -10.0580   6.0264
#>    2.3149   0.1859  -1.1973  -1.8420   3.2197   3.1546 -13.9894  -2.6050
#>   -7.7419  -0.5688   7.2768  -1.8440  -1.8441  -8.5155   5.1666 -14.3720
#>   -1.3855  -8.0960  -4.6404   6.3642  -7.9190  -6.4896   5.3178   9.3094
#>   -0.8519   0.8240 -10.4609  -4.2936   1.3122   6.4527  -4.8865  10.4961
#>   -1.5579   3.8100   0.7821   6.5707   9.9785  -8.3348  -3.2338  -5.3745
#>    2.8370  -1.9096   8.2983   1.6539  -3.5285   1.1061   0.4800  -0.2923
#>    3.9113  -1.1078  -2.9107  -2.3295   7.9077  -1.9018   3.4362   5.2866
#>    2.6903   2.1756 -12.6795  -4.0129  -5.1429  14.4581   1.8809  -2.3653
#>   -2.6584  -5.6728   2.1198  10.3743  15.6931  -5.4625  -3.4577   1.5777
#>    3.0069   6.5493  -2.7788  14.0265 -10.2361   8.7432  16.8044   1.6753
#>   -2.0524   4.3672   0.6989  -0.7723   3.3823  -1.8506  -1.0491  -9.3919
#>   -1.2007   6.6784   4.7546   4.2697  -5.7898   4.0221   2.4844   2.4931
#>   -0.0043  -1.1588  -3.4180   1.3060   2.8226  -3.1198   3.0106 -19.8816
#>    1.4941   8.8653   1.2755 -16.4456   0.0077   9.7234   6.1568   2.2712
#>    4.4434  -1.0079  -2.1225  -3.1316  22.6157 -11.0571   1.5621  13.8434
#>    0.2111  -3.6792   0.0394  -0.7679  -0.2882   0.1152   0.9793  14.6527
#>   -3.4327  -0.8576   2.7774   3.5034  -3.3591  16.9762   2.7505 -16.5181
#>   -3.8671   3.5553   3.0822  -1.7639   1.9482   6.7139 -21.7831 -16.2731
#>    3.2096  -2.8426  -2.4482  -0.3423   3.8726  13.9825  -1.1746  -0.8905
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]