Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.5270  -7.0775  -3.3977  -7.4352   2.4348  -8.3649   8.5437   5.1405
#>   -0.8388   7.7905  -4.4978   4.6710  -2.5873   8.2706   8.4023 -10.3053
#>   -3.0990   2.5696  -8.8826   1.4196  17.5221 -15.2985  -5.4803 -12.3313
#>    4.1512   3.6035  -1.5476   4.7165 -11.3318   5.6042 -15.0269 -11.4169
#>    2.4734   6.4677   6.5333  23.3223  22.9951   7.9747   1.0827   7.7761
#>   -2.1367  -6.8486  -2.7589 -14.4450   6.1898  -0.0288   0.1139 -16.5065
#>    4.2686  -6.3364  -0.4480   6.4479   3.1583  -3.3021  -6.3619 -17.8473
#>    2.8914  -4.9494   7.7366  10.9763   0.1329   0.5577  -4.6853   9.2642
#>   -7.0425   2.4199  10.9861  -6.3279   3.7445 -13.0675   8.2753  13.4437
#>    2.6602   3.0611   3.9938   1.1118  -9.8362   2.5018   8.4112  -5.6263
#>    1.8775  -8.1712   4.2138  19.6712   3.7628   1.0814   5.3628  -0.7777
#>    5.7934   2.4187  -6.5286   1.8486   5.8114  -0.2766   9.0914 -15.8553
#>    1.6419   1.0965   0.8051   4.2989  12.3664   3.1115   2.6127  15.0420
#>    4.1113  -2.9806  -0.9437  -8.3337  17.9371  10.4759  -0.7272   3.1476
#>    4.8102 -20.6234   1.1208   2.4092  -1.0314  12.1568  -2.9797  -3.4876
#>   -0.5490  -6.5500   2.3715  -9.9055   0.6312   4.3020   6.4343  11.2269
#>    1.5469  -4.9121   9.0851   7.3693   1.3213  -3.5668  -7.0009   8.3837
#>   -0.7497 -11.0126  11.2468 -11.1568  -4.5983  -6.7126  -0.8847  11.6397
#>   -4.9883   9.8092  -8.0150  -7.2169  -9.0412 -10.5697   0.0540  -5.7515
#>    3.6469  -1.3457  -1.0332   4.8937  14.2570  10.2313  -2.6914  -7.2921
#>    1.4679 -18.9057  -7.9915  -3.5962  -3.6266  -7.6341   5.1456   2.3919
#>    2.9564   5.9590  12.8844  -4.7026  -7.3530   7.2336  -2.2225  13.1012
#>    6.6103 -14.1853  13.6412   3.0251  11.2512  -5.4737  -2.9957  -5.2036
#>    4.7825  -1.1209   4.0406   4.3986  -2.6125   2.0682   1.8026  -0.0250
#>   -1.2441   0.4954 -14.0447  -9.5076 -10.8333  -2.4844  -2.5022 -13.1781
#>    5.4591  -2.2643  -2.2713   9.6267 -20.2882  -6.2236  -1.9800  27.8496
#>    5.2930  -4.2767   4.4345  -2.5714   5.9144  -0.1380   2.0169  -2.1101
#>    3.0948  -8.7100   0.0678 -10.5769  -5.2841 -29.8628 -18.2096 -10.5447
#>   -0.0122   4.1644  -2.1233   7.5277  -1.1822  10.4904  -3.5346   3.9308
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]