Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.7885  -2.2801  -1.0568  -0.1952   4.2147  -7.2285   0.7629  -3.2058
#>   -3.4727  -0.1408   5.0490   0.5147  -2.9308 -18.1730  -1.9298   8.5092
#>   -1.5900   6.2075  12.5011   5.9492   6.2536   0.0712  -2.6906  10.2276
#>   -7.3466  -3.7904  -5.2395   3.6772   1.2757  -5.1705  -0.4176  12.4031
#>   -2.1433   1.1754  -1.9916  -8.0261   1.2584  11.3444  -4.8073  10.9414
#>   -6.3095   1.0929  -0.4715  -4.9676  -3.4039  -8.3857 -11.7641   2.4431
#>    1.3937   0.1004  -5.8268  -9.8049   5.9107  16.7096  -7.5088 -18.4195
#>    5.4200   9.2689   5.4037 -14.5121  -1.2932  -8.2488  -6.7390   4.4917
#>    3.5890  -1.5542  -7.0352  -6.9025  -0.1340   3.2514  -7.2867 -11.6339
#>   -8.0005 -10.6174   1.8372   1.9296   5.7832  -0.4257  -1.9858   2.6988
#>    1.1476  -2.0578   0.5525  -1.1695   0.2035  14.4854   4.7443  -4.0484
#>   -2.9160   3.2977  -5.0178  -6.4556   8.0641  10.0457   6.1282  -7.8803
#>    1.6399   0.8166   0.4827   1.1067  -1.5868   7.5116  -3.3303  -0.7155
#>   -7.9887   3.2727  -3.6145  -7.6098  -1.6678  -0.6575 -18.1793   5.4333
#>    7.7714   4.7138  -8.3306  -2.3688  -6.1978  -1.1429 -15.0474   5.4664
#>   -0.1195  -2.8703   8.5013  13.3155  -0.2399 -14.2894  -6.1813   6.0880
#>   -8.7201 -12.8916   2.1662   6.7200   9.4547   3.0067  -6.7131  11.3881
#>   -0.4037   1.1522  -1.4588   1.5636 -12.4776   1.4595  -4.2072  -0.8486
#>    1.9305  -1.9970   2.1268  14.2814  21.0858   0.2863   5.0153  -5.5245
#>   -9.6288  -3.5877  -4.8579   3.2047   6.6881   6.1188  -1.6789   7.5985
#>    0.9459  -2.0819   7.1965  -6.2950   0.8004   6.7304 -12.1986  -8.4724
#>    0.1653  -3.6657  -9.4902  -0.7605   8.8170   2.6183  -4.0468  -6.0197
#>   -4.4629  -2.5473   6.4484  16.0554  14.2688   8.5254  -3.4299   4.4903
#>    0.6526   1.6856  -4.5978  -6.7692  -7.1941   8.3454  -1.4451   9.6361
#>    6.0617   4.5319  -7.6559 -15.0337  -8.1645 -12.0557   9.6490  -9.1727
#>    2.6597  -2.3284   0.7281  -9.6219  -2.7178   8.3553   5.3211  -4.2126
#>    5.6901   4.2809  -8.4012  -5.6097   5.8341   2.4902  -3.1227   5.4798
#>   -8.5134  -6.7098   9.4376   9.6477  -8.5639  -0.8178 -12.3633  13.1421
#>    0.4605  -3.1295  -7.7866  -1.8716  -8.2125   0.7881  16.6149  -5.8093
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]