Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.5237  10.0624   6.5696  -9.6285 -17.6108  -9.0355  -6.4312  -9.8325
#>   -0.5044   3.2957  -4.5875  -1.4871   1.5551   1.9254   6.9079  -0.2327
#>    2.5011  -3.5853  -7.2950  -6.8720  13.4926   1.6608  -0.4208   9.2924
#>    1.5863  11.7164  -2.5520  10.8047 -11.2887 -18.1660   9.6086  -1.4361
#>   -2.4367  -7.3267  -2.1425   1.8565  10.4976  10.3598 -18.7527   8.2405
#>    6.8331  -4.2810 -18.5461   1.3136  -8.3142  -3.5306  -3.2182   6.5688
#>   -4.1092   0.5153  -0.2940  10.0087   1.7067   0.4355  -4.4560  -2.9591
#>    3.8147  -8.6655  -5.6381  -2.2784 -10.4453   4.5833   8.6060  -2.5653
#>   -0.0491  -7.1790  -9.4918  -8.6849 -12.7260   3.7098  -3.8582   4.0636
#>    5.7397   5.9462  -3.5822   9.6950   4.2602  -5.8133 -10.5256 -16.4495
#>   -2.4393  -6.0522  -0.6689  13.9573   4.9405  -6.7126   9.4100  -0.1913
#>   -4.6010  -7.5554   5.7197  -2.1217   6.7539   4.3917   2.6836  -5.5650
#>   -2.0742  -0.2034  -0.9199   0.3246   0.6154   6.2289 -10.3348 -16.4070
#>    3.1987   5.9464   6.0459  -8.0339  -1.2188  -3.8908  10.8476  -5.0452
#>    1.5962  -7.2915   6.9144   6.8274  -5.5776   3.3359  -3.7902 -17.2546
#>   -1.1370  -8.5538  -2.0553  -5.5052  -6.3423   6.0585  -9.4456  16.0721
#>    0.1607  -1.1983  11.1766   0.1005  -3.5057  15.9107  -1.4845  -9.5371
#>    0.7599  -6.7298  -6.9932 -10.8972  -5.5760   2.4340 -11.1141  -0.5354
#>    4.8027  -7.8931   0.0184   3.7582  -8.7712 -14.8955  -5.7138   2.6668
#>   -2.1438  14.3435   6.4167  -5.9340  -1.9317   1.4968  -9.3316   9.7009
#>    1.4446  -0.3532 -12.2853  -9.8357  19.8976  -9.0295   6.9103   6.8389
#>   -0.2995  -2.6250 -12.0607 -11.9265   0.0316   5.5599 -13.0726  -7.0834
#>    0.7716  10.2410   5.5622  -8.0517   5.8905   4.6031   4.8129  -0.9487
#>    1.0111   0.9532   8.4136   8.7461  28.2909  -6.8504   2.2409  12.6339
#>   -1.1989 -13.1445  -3.4895   1.5887 -13.3787  -0.9308 -14.4712 -12.7314
#>   -1.5363  -7.8159   0.9757  -0.4798  12.7698  -7.8681  -0.8593   5.9529
#>    1.7205   0.0789  11.4384   6.8153  -7.4257  -3.1484   4.6512  -7.1322
#>   -1.8929   1.8363  -5.3713   0.3682  -2.7687  11.3146  -5.5234   0.1609
#>   -0.0963   0.1787   6.2564  -2.0459   3.5667   1.5860   3.8154  -3.3143
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]