Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.9097 -10.8449  -0.5598  -5.6137  11.6630  -6.7941 -11.4844   6.4263
#>   -0.2622  -7.4325   5.1361   1.1177  -9.1313  -8.4525   9.4743   9.4603
#>    1.4807   8.1142  -4.7125   3.5456  -1.0634 -10.2152  -1.2251  -0.0587
#>   -0.1290  -0.4860  -0.7113  -2.1552   0.6909   3.7851  10.0900  -4.5985
#>    0.9077   6.8948  -1.1827  -4.5574   2.6707   2.2599  -8.0899  -4.1876
#>   -1.4026  -2.1818   6.7642   1.9473   7.8085  -2.5072  -6.2092  -0.6727
#>    1.7120   6.0715 -11.0482  -3.9775  -3.9975  13.5285   4.9124  -4.9050
#>    1.1864  -3.5560   7.8084  -1.0366  -7.2243 -11.2726   0.8464   1.7672
#>   -3.0910   0.7909   1.3351   1.5645  13.4368  10.2449   7.5912   9.8527
#>    1.4202  -3.1858   0.8221  15.9448  10.4610   6.6395  16.3269 -21.0941
#>   -1.2459   2.8529  -2.9565  11.0058  -5.5218  -8.4319 -10.1694 -11.5533
#>    2.8767   6.5808  -3.4684  -3.2660   9.6074  -1.7131 -12.1926  -9.3052
#>   -1.3886  -2.9162   0.2542   0.5907  14.3194   7.5376   0.7822 -17.3837
#>    1.0428   2.3276  -2.7999  -8.2613 -12.4584  -2.3533  -0.8122   1.4959
#>   -3.9125  -5.4291  -0.2230   4.3619  10.7084  -0.7624  -5.4095  -0.6987
#>   -3.8640   4.5615   0.8236   6.2086  -0.0789   9.8067  -8.4822  -7.3934
#>    3.9556  -0.4930  -2.7833  -5.0925   4.4315  -2.1108  -1.9154  -4.9371
#>    2.6477   5.3480  -2.5462 -10.8057  -6.3930   3.1495  -0.2752 -18.1114
#>    0.8224  -3.2485  -1.8686   5.1701   1.4988   2.0885  12.6249   7.8901
#>   -3.5613   0.9049   1.6825   9.8236   8.1214  -8.9627   3.6348  18.0382
#>    1.3397  -3.1609   3.2106  -8.5108   4.8412 -12.6662   7.5864   3.5720
#>   -3.1793   5.6117   2.1614   4.1646 -13.8851   1.5751 -10.7289  -4.7889
#>   -0.6981  -3.1896   2.0673  -0.5260 -10.3275  -5.5897   2.8279  -8.7245
#>   -0.3592   5.6249  -2.1499  -1.3810   1.3194   5.4440  -0.3058  11.4277
#>    3.7578   1.4550  -1.9959  -4.0722  -6.6084 -12.4928   5.7697   4.3591
#>   -1.3567  -4.1056   0.0870 -14.2934   2.2818   9.6693  -0.3575 -11.2534
#>   -4.8673   1.9884   2.5257   4.3107   1.2375  -0.7313  -7.4188 -20.8097
#>   -1.6082   2.7254   3.6646  -2.5130  -6.4680   6.0188  -9.7025  11.0903
#>   -3.5238  -5.9525   7.7225  -2.5060   2.8654  -3.4206   7.0542   5.6902
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]