Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.7505   5.0060  12.5356  -0.2791   6.0064  13.4863   0.0072  -0.6845
#>   -1.1451   5.9248   2.5302   8.5739  13.5796  -3.4128  -2.7278  -4.3850
#>   -2.8486  -2.4749  -0.3857   2.7689   7.1173   4.2884  15.9678   3.8454
#>   -2.6466   0.7118   0.3041  -1.8161   3.6762  -4.2768   7.5757   7.8010
#>    3.4405   4.4147  -2.6040 -14.3663  -8.6128 -12.2776  -8.2903  -0.9419
#>    6.5478   1.8454  -2.8304   5.6031   5.6114  15.3512  10.2569   3.2274
#>    0.9365   2.2101   8.6700   7.1336  -1.7919   6.0507  -1.2271   0.4719
#>    0.2081  -4.3623  -1.0232  -2.6298   1.4255  -0.4024  -6.8290   9.7110
#>    0.7049   0.6957  -8.5644  -8.5988  10.8865  -2.8040   3.0206  -5.2042
#>   -1.3341   2.0939   1.2839   8.5186 -10.3528   5.8007  -2.5179   4.5116
#>   -2.6022   5.6386   2.5202  -2.4361  -8.7141  -2.6191   5.7789   7.5002
#>    2.4012   4.2604   9.9135  13.2660   7.3072   3.3891   8.5205  12.6828
#>   -1.8997   0.3139  -4.6000  -6.6094   6.5159  -7.3040   4.5132   5.5845
#>    5.0698  -6.1426  -7.4831   4.0397   2.9905  -1.8961   4.2504   5.5717
#>    0.2980  -0.9066   1.1899  -7.1173  -0.8047  -7.4866  -9.5985   1.4256
#>   -1.6808  -2.7766  -4.3823   7.5833 -12.1792  15.7383 -12.4979  11.6062
#>    0.7077  -8.4456  -2.5742  -0.1224   1.8622 -20.0753  -3.4527   2.1727
#>   -8.6154  -8.1361   2.4082   1.5814  -1.7564 -16.5906   3.2341  -4.6574
#>   -5.6484  -2.9125  11.8694   0.2483   5.5733 -10.7060  -5.0723  -3.8352
#>    1.7926  -4.8797 -17.7580  -2.8978 -11.8893  10.6123  -4.2875  16.8078
#>    0.7859  -5.2310  -4.1242  -3.8988  11.0590  -6.9978  -4.7155   7.2012
#>    0.0933   6.0687   3.1595  11.7073   0.0132  -3.8246  16.9893 -13.8540
#>   -0.2237  -5.1499  -0.0143   5.3429  -4.5665  -0.2389   6.2122   6.1877
#>    0.5437  -2.9433   2.0035  -8.3958 -14.8706 -14.8623  -8.2705  -5.1712
#>    1.8055  -4.1734   3.4819   2.5331  15.5654  -2.1504  -0.2492  -4.7158
#>    3.1670  -1.5847  13.9077   1.7360   2.0977  -2.2288  10.3535 -15.5558
#>   -6.1221  -4.4462   0.3232  14.5832  -0.0941  -4.0537  -2.1446  14.2839
#>    0.2191  -5.0792  -8.6849  -1.1365  -1.1111   6.8776   0.3456   2.0647
#>    1.2096  -0.3119  16.4523   0.6678  -1.0404   3.8805   3.7715  -4.4779
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]