Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.8799   0.1709  -5.0043  -1.7902   0.6865   4.7132   4.4754   6.7609
#>    1.8325   0.6621   8.6633   9.5534  -0.4690  -9.4043   9.3160  -3.5930
#>   -1.2293  -0.0526 -12.4134   1.4585  -0.0448  -2.7811  -1.1833  12.7345
#>   -1.2900  -3.2071  -3.9128  -7.0228  -3.5607   7.7530 -11.2758  18.0507
#>   -4.5629   7.2393   2.4682   1.2435  -4.3610 -10.1721  -5.7255   4.2153
#>   -8.1370  -0.3174 -10.6116  -8.9182  12.5891   1.6483 -18.9568  -1.9208
#>    9.7618 -14.1625   1.6532   9.1842 -12.4830  -9.5472 -12.3740  -5.3309
#>   -1.8980   6.2206   3.7788 -18.7247  -3.1067 -23.0987  -1.9390  -1.8404
#>    2.7619  -0.4128  -4.9479  -8.9712 -11.0407   5.4264   1.3539   4.1874
#>    2.5328  14.2602   3.8089   0.3922  -8.7587 -23.0833  -8.5788  14.7294
#>   -5.2705  -4.6468   0.9748  -0.3222   7.0619   3.2169  -1.9975 -17.1357
#>    0.5142   5.1620  -4.6200  -9.5605  -7.1089  -8.0836  -2.0437   1.4241
#>   -3.0002   6.3930  -0.3731 -10.5418   6.9918  14.5263  20.6009  -2.3099
#>    2.2488  -2.5052   4.5472 -21.8036   4.9807 -14.6402   6.2735 -16.4638
#>   -1.6523  -1.6341  -1.1481   4.0360 -13.5154   8.5523  13.4331  -7.0147
#>    1.2298  -0.7050   3.7525  12.3075   7.9072 -21.7692   0.0975   4.2183
#>    2.1166   9.7904 -11.5644   3.6238   4.8565 -10.1249   4.9116 -17.0857
#>    3.8924  -4.2059  -2.3066  -1.6274 -18.1830  -7.7518 -16.5882 -11.2379
#>    5.5934   0.7261   0.0122   2.7165   7.1791  -7.3784   9.5669   8.9386
#>   -0.2445   1.2126  -3.9223  -3.3983  -2.4561 -18.9600   4.3493 -11.1158
#>   -3.0515   4.4385   2.1865   7.3889  -1.2626   1.8553  -7.6986   9.4961
#>   -0.3704  -1.5541  -2.5272   2.5496   2.9701  -5.3007  -1.3949   1.7024
#>    3.5959  -5.1109   5.2029 -11.6581  -4.3013  -6.9595  -6.8140  -4.0367
#>   -2.2924   1.3937 -10.2941   0.1228 -13.1381  -8.5471   3.9104  -6.0945
#>    0.7766  -0.8445   3.8277  -2.8501 -10.2973  -9.2201  -7.4022   1.9682
#>    3.9057  -3.2279  -5.9142  12.9125   9.7427 -12.4114  -4.5902  -7.0528
#>    4.6664  -3.7564  -6.4876   3.2044   2.1107 -18.2819  -6.4801 -15.3826
#>    1.2794  -2.8215  -6.5513   6.9946  -7.6311   5.6371  -9.2182  -1.0262
#>   -1.5278  -8.1466   2.5251   0.1299   0.4965   2.2994   0.5450  -0.0952
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]