Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.4069  -0.1197   4.0980   1.6150  15.0847 -16.4871  -3.1373  10.1927
#>   -0.7841   1.2042   4.4430   6.2130   4.9037   4.7507   7.3810 -10.8770
#>   -1.9036   4.4582  -6.8272   4.8734  17.1275   9.4892  15.4601  -1.8450
#>   -2.3809  -2.7829  -0.3495   5.1555  -3.6501   3.8030  -5.5284   3.0688
#>    1.3363  -1.4773   2.0865   5.0251  -6.3022  -5.6681  -4.0481  -7.6098
#>    0.7942   2.2171  -3.5640   2.1501  -5.0644  16.0372  10.0839   1.7582
#>    0.7792  -7.1515   8.0799  -4.6220  23.8168   0.8403   5.5185 -10.1932
#>    2.5026 -11.3390  -0.2043  -2.4816  -8.4310  11.9136  -0.5884   1.4936
#>    3.5427  -0.9685   1.2853 -10.9065   0.6485   2.7390   8.2589 -10.2406
#>   -1.3993  -8.5720  -0.1651  -0.5086   3.0432  17.0533  -1.4336  -0.3670
#>   -0.3936  -6.9102  -2.4992   6.0883  -9.6585  -3.2839  16.9078  -9.5904
#>   -0.2859   1.7620   2.1830  -0.9866   4.9171  -1.3645   4.3221  -9.1749
#>    1.1528  -8.1995   5.0328  -6.9427  -2.6766  -3.9927  -1.1029 -17.0190
#>    0.1034   1.1547  -2.7194  -5.7404   0.1929  -3.5150  -9.4110  -2.4269
#>   -3.2740   5.3992   4.6644   9.3624   2.4961  15.8828   0.3160  -4.1433
#>    1.2795  -2.1974   5.0801  -8.9178  -8.2434  -1.5555  -1.5127   0.3653
#>    3.0517  10.1797   9.5428  -0.7965  -5.2860  -4.5492  -7.8347 -15.2006
#>    1.7936  -1.4818  -4.9570  11.9719   7.9283  -5.7599 -13.3461   3.4196
#>   -4.9861  -0.3481   0.0570   1.3319  -4.1974  -6.2131 -11.7402   3.2459
#>    1.3096  -1.9882  -0.4437  -3.4903   0.5792   5.8393   8.8431   5.8408
#>   -0.3872   5.9574  -1.2303   9.1596   1.7397  14.8439  15.0838  -7.7838
#>    1.3020   1.5453  -0.3932   4.6907   2.8174   6.4073   3.3764   4.4172
#>    2.1812   5.1248   2.6330   1.1824  -6.9027 -19.4734   8.8413  -0.5567
#>   -0.1983  -5.7500  -5.7003   0.6490   6.9894  -3.6505 -12.4637  10.0812
#>    1.8914   3.3599  -7.1709  -9.1767  -2.0214 -14.3194  -4.7224 -17.6773
#>    0.2040  -3.6341  -2.1130   5.0341  -1.3379  11.3237  -9.8964   5.7207
#>   -1.6305  -1.4897  -7.9106  21.0599 -13.6908  -1.4958  -4.3441  -2.9578
#>   -0.2567  -4.8671  -7.5389  -3.7392  12.0097   6.0363  -4.5523 -10.4714
#>    1.7879   6.0604   0.6410 -12.9689  -0.4388   5.2911   9.5102 -14.4172
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]