Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.7689   4.6475  -1.7674   9.3149  -1.3619  -5.6118   2.1947  -9.7455
#>   -2.7404  14.9670  -4.8592  -0.1503   2.6079   1.5498   6.7719   7.5383
#>    1.6618  -0.1011 -11.7264 -11.0958  -8.5130   4.0845  -4.8909 -20.5340
#>    6.3209  13.1526  10.0108   9.5727  -0.9207  -5.9028   1.0896  -7.7156
#>    0.8836  -4.4208  -9.9026   4.4050   3.2276  -6.7266 -12.2824  -9.0129
#>   -2.0832  11.1071  18.1879  -0.9746   8.2685  14.5936 -10.8310   1.7901
#>    3.6747  -7.8070  12.2712  -3.8776  -2.8487 -14.2964  13.1777  -4.4876
#>   -1.0323   3.8874  -7.9357  -8.1663 -11.3608  26.2493 -11.5892   0.4342
#>    1.7242   0.1978  -4.2882   2.3100  -0.9075  -4.7114  16.0722  -0.2231
#>    2.4909  -7.4884  -1.0353  -0.6351 -14.8978  -3.9377   9.8624  17.8450
#>   -2.0715   5.1165  -0.0892  -2.6473  -7.1820   9.2841  -6.6683   4.6218
#>   -2.0002  13.5715   2.6808  -9.4669  -2.7662  11.4939  -2.3671  -3.4857
#>    4.7936   1.5516  -2.4055   3.4234  10.1382  -2.4677   7.4992 -16.8683
#>    2.8344   1.8802  -1.4124   4.5700   6.1872 -14.1071  24.6518 -16.0829
#>    0.4600   1.4927  -5.8845   5.3398  -8.6822  -8.4723   4.7361  -7.9548
#>   -0.8660  -2.3055  -3.5457   1.0373  20.8191   3.0194  -2.8713  -2.4312
#>    2.6946  -7.5134  -0.1227   7.5370   6.3756   5.0554 -11.2862  17.2121
#>   -7.4300  13.4017   2.9366 -22.5878 -14.2184  24.2076   1.3362   4.9983
#>   -3.8092   1.3443   5.1823  -2.9968   8.5017   6.4769   1.0950 -11.2984
#>    0.6533 -10.2092   1.7800   2.2606  -2.5358   0.8724   7.2974   5.5999
#>   -9.7113  -0.1799   8.9473  -9.6579  -6.8365   4.0507   2.4847   7.9873
#>   -0.0523  -0.3161  -0.0223   7.9080   4.5211 -10.0830 -14.8304   5.9363
#>    1.9103   1.8842 -21.8326   2.0177  -5.8141  -9.7039 -12.1882   3.5653
#>    3.4102   4.8462  -1.7762   8.9625  -5.5982 -11.1444  -7.8221   1.6117
#>   -0.9271 -13.1409   2.5832  14.6223 -10.8653 -17.5214   1.6301   7.6818
#>   -0.7838  -8.3291  -8.5036  -2.7904  10.3909  -2.6425  -7.6529  12.1349
#>   -2.5828   1.3445   6.5737   9.6467   4.4421   0.9979   2.6118   3.3873
#>    3.4029  -2.7927   0.2043   6.2720  -6.7646  12.9315  10.8672  -8.5663
#>   -3.1121   3.4341   8.6281  -5.5004  -3.2214  -5.6487 -22.7876   4.1840
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]