Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.5367  -2.4599   0.3673  -3.2991   7.7544   6.0189  -5.1632   5.3884
#>    3.7506  -5.0329   6.1436  -8.7189  -6.4242  -1.4465   7.0004  15.3948
#>    4.2100   3.4829  -4.7300  -8.3943   6.3414   5.4047  -0.2449  -6.7681
#>    0.5880  10.2471  -3.8930   3.2850  -4.1049  -1.8516  -1.0579 -16.5007
#>    0.5682  -2.8133  -3.1838  -5.0660  11.3649   7.4369  -9.0168 -14.0571
#>   -1.3161   4.2581   5.7918   4.7652  -4.6826 -10.9813   2.9330  13.2872
#>    2.8677  -6.6767  12.0127  -1.5561  16.1395  -1.0055  -9.6532 -13.9119
#>   -3.6142  -1.3780   4.4886  -6.8984 -11.8240   8.9689 -12.8392  17.0830
#>    2.0466  -2.3147   1.1056  -3.3212   5.4463  -4.0506   2.6849   4.6656
#>   -1.1089   0.0149 -11.8695   1.2780   4.7120  10.3663 -10.5054 -10.0153
#>    1.7265  -9.5947   4.3731  12.6004 -19.7015   8.8716   8.1629  -0.7449
#>    4.8869  -0.7877  -1.0494  -2.8024  -3.5489  -3.8731  -7.9813  -6.4160
#>    4.3013  -8.8125  -0.7444  -0.8537   5.0397  21.0182  11.7622   3.7723
#>    0.7381   4.0858 -10.5694   5.3231   0.6153  -1.0041  25.8577   6.2611
#>    6.7833   4.6451 -11.3513   0.3569  -8.8155  10.0966 -10.7739 -20.8554
#>   -3.6558   2.0802  -3.2777   0.7962   8.7877 -15.1614  -9.2554  14.1654
#>   -3.1936   4.6323  -0.1064  10.3787  -8.0796  -7.2026   7.9008   9.3188
#>   -2.6941  -3.3800  -2.2278   3.9827  11.2911  -9.2107 -13.2026  -0.3604
#>    2.0100 -12.8576   3.8383   5.6011  -8.1405  -5.1636   6.0256  -6.2834
#>    1.9924  -9.7734   8.6251   3.6029  -3.1020  -1.3843   5.3821  15.8041
#>    7.9364  -2.1465   2.5645  -8.9198   9.0361   6.4870  -0.8415  -6.0604
#>   -0.0597   8.5100  -9.9643   3.1255   1.4667   4.8084   4.5627   3.8067
#>   -0.1319   4.9886   1.3872 -16.2552  10.2060  -3.4334   9.2045  -9.1966
#>    1.1465  -3.3071   6.4994  -2.1236   7.6571  11.1202 -11.0814   0.7627
#>   -1.0877   4.0334   4.6451   2.8563   0.3372  11.0672  -7.7018  -6.6973
#>   -2.1790   2.0566  -2.4895  17.0219  -2.4190  -8.7999 -11.4482  11.7921
#>   -1.7531  -6.0764  -4.3974   3.4277   2.2513  13.4468 -12.0790 -11.1230
#>    4.1762 -10.9117   1.2307  -2.2950  14.7487 -13.6805   2.3339   5.8152
#>    6.3354   0.4142  -6.6965   6.9858 -14.1094   9.2136  -6.1713   1.1442
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]