Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.4040  -4.7670   2.8984  14.0785  -2.6570   4.9906 -11.5009   2.1073
#>    1.2121  13.3591  -3.8887   3.1041  20.4154   6.5012  -8.0308  -4.3165
#>    4.8815   4.8411   1.0086 -19.4966   0.5809   5.3445 -13.5418   5.1375
#>    4.4182   7.8989  -5.8321   8.6430  12.3323   2.7910   1.5169  -1.1778
#>   -6.0048 -17.1189  -3.8135  -5.3398  -3.3984  -0.5244  -0.1203   9.4461
#>   -2.0518   7.4809  11.2211   7.0271  -0.1210   3.4569   1.6119 -13.8011
#>   -0.4394   0.0956 -15.5277 -12.3243  -6.5231  -5.2635  13.0179  -2.3371
#>    6.3403   1.9428   5.1838   8.5344  13.9260   4.2712   4.7304  -3.0855
#>    2.0725   0.5299 -13.1944  -0.7928   0.3713   7.0517  16.0894   5.9227
#>   -5.4046   3.4357   1.3492  -0.5511   8.5674  -9.3009  -1.2144   7.4328
#>    3.9340 -18.6490  -1.3590   8.5217  -1.0457  -9.4745  -8.2894  20.1417
#>   -0.3253  -3.4148  -7.1466   3.5443  24.4860  19.0038   2.4816  16.5127
#>   -6.5822  -2.8574  14.6445   0.7901 -10.2574  -8.0443   3.0723   3.3498
#>   -4.5627  -2.6209  12.2289 -10.5901  -4.0771   8.1428  -6.7255  -1.9538
#>    4.7866  -0.3806  -0.3954  -5.4798 -12.2911  -0.9298  -4.1606  -3.7346
#>    1.3555  -3.3733   0.0277  13.7918  -2.7325 -17.9441  -6.0257  -5.3833
#>  -10.2142 -11.5412   0.6102 -17.2555  12.5374   3.5899 -12.6239   7.1256
#>    1.0447   9.5627  -8.0902  -0.1268   1.5138 -16.3204  -1.0944 -14.3710
#>   -7.1281  -3.4297   2.0821   1.0648   3.2234  12.8603   2.2830  -3.9693
#>    1.5350  -3.1445   8.3040  -0.6454  -2.3678  12.5321 -13.0146   6.7917
#>   -0.8120 -11.7403   2.2473   2.6640 -17.8841   2.7409   0.4991  17.4615
#>    1.4020  -4.5279  -3.7367   2.5591 -15.1103   0.9769   4.4842  -8.3560
#>   -5.0568  -4.7409   0.4295   2.3539  -6.3404   0.6506 -17.9345  15.3424
#>   -2.7891  -1.3161   3.2285  -6.8561  11.5297   0.4020   1.7666 -10.3543
#>    5.1991   5.3348 -11.5357  -6.3492   4.2246  -9.7453   8.4437  -5.4266
#>    8.0111   6.6930  -8.8121  -7.6040  -6.8828   8.5966  -5.6453   3.3012
#>   -7.0201   8.1721   3.5787  -6.6211   5.0907   9.3925   1.5651   4.4523
#>   -0.6124  -2.4701  -4.5279  -5.0817   3.3388  -1.4714   1.5793  -8.7079
#>    0.1941  -5.2159  -3.6042  -8.5560   0.5739   5.4191 -18.4476  -9.3574
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]