Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.5554   4.9461   4.1070  -2.9209  12.1293  -7.8611   6.1886  -3.9430
#>   -0.8723   5.0557  -8.0385  -3.5759 -24.5148  -7.6386  -6.6280  10.4968
#>   -1.2920   1.2621   0.7798   2.8229  -1.8434  -3.6983 -11.7598   4.7367
#>    1.7523  13.9413   2.9849 -10.3468 -14.8812  -8.9262 -13.2649  -7.2913
#>    1.7877   6.0942  -1.0228  -0.8429   6.6626   4.1997  -3.5775  -4.0136
#>    7.8763   0.0193   0.0980  -8.5628  -5.2812 -22.0899  -2.8366  -4.7341
#>   -2.0483  -1.0382  -8.3528   3.4897 -10.3400   5.4630   1.1989  -4.2868
#>    4.8792  -0.0176   3.7372   8.7884   4.4590  -3.9471  11.6170  11.4420
#>    1.0621  -7.7365  -3.9382  -2.3343  12.0063  -1.3668   8.4252  -3.2067
#>   -2.0101   0.3015   6.0610  -1.2422   6.9192   1.6646   4.5083  -4.1192
#>   -1.4118  -0.4497  -2.4581  -3.5519   0.1312  -6.4158   8.1474   0.7794
#>   -1.8649  -6.1772  -5.5359 -12.9749  -5.1065   3.9136  -7.5851 -10.5036
#>   -0.2047   4.2196   3.0084  -2.5480   2.2089   3.2042  16.5366  -5.1880
#>    4.5223   2.2018  -9.7674  -1.3570  -7.1713   3.6485  -2.3311  11.5017
#>   -1.1445  -5.8797  -2.6039  -2.3035   7.5157   8.8014  -4.6716  -7.3388
#>   -0.8740  -2.1686  -0.0763  -9.3552  -3.4608   0.3498   1.7516 -13.8121
#>    4.0002   0.5019 -10.1652   4.2704  -6.6942  -2.9129   6.6402  -0.2023
#>   -1.5542  -0.4812  -2.1348  -2.2164   0.8008  -4.2290   4.8832  -3.9264
#>   -6.8564  -1.8204   1.0509   4.2053  -6.3487  12.8621 -11.3432  16.5921
#>    3.5090   6.6001   0.9110  -8.1427  -5.1442  -8.5463  -6.2839   9.1890
#>    3.3198   7.0557   9.8957   1.9285  -7.7751  -1.5889 -14.5673  -3.2656
#>    3.8736  -4.7768  -0.1486   1.2798  -7.1301   6.4778   1.6638  -3.7370
#>   -0.2313   3.0470   2.2824  14.6141   0.1944  -3.7858   8.9947 -11.2622
#>    1.2369   0.4900  10.4096  -8.9306  -4.1675  -4.2864   6.6003  -7.2427
#>   -2.2399  -1.3731  -0.9625  -3.6695  -7.4776   1.6245   9.4635   7.2081
#>   -2.1825  -1.1918  -0.4207   1.1540  -3.9700   3.0978  20.8432   5.9620
#>    0.8689  -0.9751   8.4464  -4.0913  -1.7655   7.5836  -1.3001  -5.7690
#>   -5.6145  -4.3475 -14.4099  -1.7247   3.4855  -5.2808   6.9556   5.2752
#>    0.0241   6.9958  -4.2873   2.2318  -0.7690   2.6369  -3.9919   1.8477
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]