Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.7679   1.8475  12.6594  -0.5208   8.6694  -9.8487  12.5371  -5.6304
#>   -2.6567   2.6873   7.0529   6.0565  -3.4627  -1.9038 -17.6429  -1.1574
#>   -6.9569  -1.8998   1.2192  -6.3967  -7.7204   3.5360   4.0646   2.5719
#>   -4.0607   2.9177  -3.3777  -5.4775  -7.6988  -9.6618   2.5833  13.0737
#>    2.1684  -1.8758   3.2062   4.1880   1.1992   2.0858   9.3372  -6.1616
#>    0.5791  -3.2319   3.1285  -5.5478   0.9292 -19.6704  12.0121   0.6958
#>   -2.4310   3.2710  18.8541  -8.9657   0.8398   5.8747   8.6369 -15.0763
#>    3.9358   2.4561  -8.1689  -3.6101   9.2910 -14.5597  17.2045  11.7462
#>    3.0240  12.9214  -3.2527   3.0553   7.7271  10.4085   4.9859  -7.2331
#>   -5.7196   4.2001   0.7593  -1.5567  -3.3662   6.2998   4.6939 -13.2123
#>    7.1434  -3.0394   0.1588 -13.4222   5.0676  -1.4284   6.3350   3.7440
#>   -4.5835  -3.0484  -4.3214  -1.0896  -4.9608  -3.3479  13.7131   7.3636
#>    2.4456   0.9423   8.3306  -8.5093   9.8207   4.4933 -16.5815  16.7232
#>  -14.3247   0.6350  -7.9660  -8.3518  -7.2348   1.8385   0.8973  -7.5744
#>   -5.8376  -3.0038  -5.6967   0.8281  -6.1611  17.0551 -13.3257  -2.7077
#>    1.6363   1.4369  -1.7048   6.9553   7.6841  -2.2117  -2.2668 -15.3990
#>   -3.1654   1.9216   0.0707  10.0099  -5.2775  14.0722  14.5881   6.2666
#>    3.2252   3.0309  -1.9087   2.1216   8.0837  -5.7676  -3.6972  -5.3935
#>    0.5379  -4.4463  -0.5101  -7.8706   7.8016   8.1425   6.5242  14.2772
#>   -0.4414  -2.2324  -5.7206   1.6147  -4.4671   6.0683   2.1945   3.1285
#>    3.1606   3.9385   2.4503  -3.9066  -6.9867   6.1809 -23.7611  13.8367
#>    0.8496  -0.9425   5.7031 -10.6927   4.5218   2.2541 -16.7052   0.8395
#>   -2.1808   5.0136  -1.1644   0.6752  -9.3742 -14.5690   2.6184  -8.3037
#>    0.6134   8.1183   1.9706   5.7830  -4.2162  -4.9556   0.5804 -15.8950
#>   -2.4661  -1.9710   8.7157  -4.7267   4.1761   7.3852   4.4208 -16.2231
#>   -5.2591  -2.9483   7.0672 -13.2658  12.8751  -3.0017  14.2992  -4.6200
#>  -11.9531  -4.3700  -1.2765  10.7900   0.6958  -5.3432  12.3157 -14.5925
#>   -1.1064   0.9041  -5.0938   7.1425  -4.5799   4.8644  -2.3410   7.6048
#>    2.2796   9.8378  -1.6771  -3.2151   8.2782   1.3704  -9.5617  -0.4075
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]