Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.1410  -9.6668   2.1398   1.0053  -4.7530 -14.0257   2.7310   2.4351
#>    0.8398   9.9952   8.5242  -3.2726  -3.2987  -1.1045   5.3089   6.0753
#>    2.0590   0.3207  -4.7137   2.2744  -8.1181  -8.6733  -9.0343  -2.5292
#>   -4.4672   7.5143   3.5443  -3.7683  11.2123  10.1455  -4.1215 -18.3179
#>    8.2579  -1.9592   4.3935  13.2196 -18.7654 -20.3222  -3.9500   6.1021
#>    3.4422  -4.5069  -5.7204  -9.7703  -8.2040 -19.6578  -4.7326  11.5395
#>   -1.6618  -5.3167   2.7794 -13.5952  -0.0048   1.6799  -5.6761  11.6332
#>    2.3373   1.4744   0.9256  21.9309  -2.9200  -5.9901 -14.8426  -7.2350
#>    0.8261  -1.1731  -9.0414   2.4128   1.4375   0.1615 -11.0219 -14.9681
#>   -0.1795  -0.6775  -0.3629   9.6100  -5.3705   5.9546  -8.5937 -20.8982
#>   -2.0466   3.0308  -2.0754  -4.0401 -11.9605  -6.6435 -14.4583  -6.5135
#>    7.1728   0.4090   3.8976   0.7101  -6.0006  -1.4514   3.3853 -17.1444
#>   -3.5601   7.7187  -6.5101  -0.9154  -5.9266  -3.6362 -12.2633   5.4212
#>   -5.9805   0.2938   3.3727  -2.2802  13.3007  -1.2569   9.2933   2.7786
#>    1.8377   5.3580   4.2809  -2.2760  -2.6677  11.0801   9.3626   4.7208
#>   -4.4950  -0.7380  -4.9155   0.0387 -14.1537  -5.2653  -2.1777   4.9413
#>   -3.6117  -0.8063  -9.0273  -9.4965   8.3505  -6.7242  -2.3522 -10.7454
#>   -0.1559   2.1588   2.6355   2.0155  -0.7247   3.4230   5.1252 -12.9398
#>   -4.0341   4.3075  12.0209  -3.3174   2.0137  18.7905  14.9652   5.0414
#>    0.7676 -12.0396   9.6247   2.3138   6.3131  -9.9843   5.7069   5.7344
#>   -1.0263  -2.4172  -3.6921  -3.1080  -5.3970   0.5884 -10.1550   1.3175
#>    2.0183   1.8618  -0.8159  17.1191   4.7883   1.4305   7.5388  -0.8256
#>    3.1803  11.7279  -3.6809  17.9746  -4.4194   4.0441   3.5579  -6.3156
#>   -2.3518  -1.1949  -7.0610  -4.0268   8.2414  14.5396   1.7403  -4.3772
#>   -1.2209   9.0274   5.1893   4.1414  -4.5002  -9.9388  11.1008   9.8169
#>    4.3137   7.7242  -2.2343   9.5547   6.1159  -2.8463  -7.7297  -4.6633
#>   -0.3923   0.7097  -9.0398  11.6806 -16.6643 -14.8757  -2.1687  -6.2958
#>   -0.3890  -7.4957  10.9194   0.5779  14.2710  -0.6761   7.7796  -3.9827
#>   -6.0442   2.5966   2.1759  -1.9795   1.6362  -5.5892 -11.2930 -10.7109
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]