Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.5580  -6.7792   0.1203  -4.0295 -10.7784   4.9684   2.6167   8.8503
#>    0.6572  -0.5552   5.8495   9.7604  -1.3021   4.8222  -4.2054  -0.7849
#>   -2.4672  -0.9571   6.4661  -6.2710 -12.1490 -11.8252   0.0657  -6.9509
#>  -12.0365 -18.8026 -17.9147   1.3521  -7.3486  15.2759   5.6010  -2.2389
#>   -7.2462  -8.1648  -2.5300   1.6395   8.0318  -2.2611  12.6243 -17.2859
#>    4.5833   8.2970   4.7053   1.9772  -0.4923  -7.1803   7.0779 -12.8494
#>   -1.1252 -14.5286  -5.2673  17.1001   3.5962   1.7557  -6.2808  -2.8646
#>    7.7417   8.7616  -9.8625 -16.2670 -23.0363 -17.3876  -2.6620   7.8623
#>   -2.4100  -3.9392  -4.5972  -5.2654  -5.9686   5.1009   4.6002   4.7700
#>    7.0696  -1.4498 -18.7965   0.3149   7.8532   4.8279  -9.5227  13.0619
#>    5.9463   5.9028   2.4321 -10.9051  18.6908 -22.1764   0.4149   6.0828
#>    9.5753   5.1825 -10.5016  -3.0871  12.9783   5.2359  -1.3489  -4.9913
#>    2.1732   1.2869  11.7202   0.1428 -12.4336  -0.0923 -10.8520  -6.7648
#>    5.2484  23.7085 -13.6117 -12.4295   9.9848   7.0603  -7.7472  13.0903
#>    5.9039   6.3727   4.4852   0.4876   8.6478   9.8082 -22.7597   1.7274
#>    2.4426  -3.0938  -1.7497   9.7193  -7.9860  -8.4180   8.8394   9.6528
#>    3.4068   4.8277   6.6525   7.1267  -4.2078 -16.8596   7.7906   5.8691
#>    1.6622   8.1624  15.7284   0.8313  -4.1401   0.8851  -9.3169  11.5381
#>    0.4645   9.9975   5.4104   4.1308 -18.4929  13.4644   4.1056  -7.5242
#>    5.2526  -0.9238  14.1111  -7.4731   7.8517  -1.2008   3.5145  -8.5335
#>   -0.1056   2.0216  -9.3618   0.6203   2.6764  12.9217 -10.4244  17.6638
#>   -5.1238  -8.3597  -5.5638   4.7968  -2.9357  -7.8534  12.8426   1.4215
#>    8.4900  14.5474   1.7605   5.4783  -9.5824  -2.6508  -6.0091 -10.4955
#>    3.2631   4.8659  -2.4698  -6.5228   9.4161  14.0446  -3.2133  -5.2996
#>   -0.4381   2.0409   0.0794  11.1533  10.9658   1.5383   6.2376  -5.4752
#>   -1.5318   0.5489   2.1509   1.6574   4.8597   2.1238  -0.1618  10.3199
#>    4.1332  -7.8813  -3.3551   6.0286  -6.7364 -15.4471  -1.2056  -1.2819
#>   -5.8712  -5.1473 -11.8369  10.8081 -12.7617  -2.7290   2.9160  -0.8394
#>   -6.6629  -6.3087  -0.5590  -8.8182  18.5255  -3.5376  16.7560 -10.0700
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]