Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.5674   1.7213   1.4314 -10.9237  -0.3565   1.6112   8.1043  -4.1796
#>    1.7345   0.2844  15.5417  -3.6151  -6.4739  -2.3718   3.5406 -11.5448
#>   -1.2777   0.0483  -1.6725 -18.6731  11.0850  12.0609  -0.4972   1.1255
#>    2.1246   0.3463  -0.0151  -4.4851  -3.3995  13.6237   6.4636  -2.3381
#>    0.0020   4.1872  -2.4017  -6.5091 -12.3197  -9.1661   8.1944   3.0823
#>    0.7239   7.3036  12.3648   7.4257  20.2407   6.9372  -1.3876  -4.6094
#>    6.1159   7.1211  -1.6314   4.6525 -15.1593  -0.5340  -3.6138  -9.3575
#>    5.2464   8.2176   0.5141   5.5222   3.9251   3.3640   7.1105   8.3742
#>   -3.8890   3.0341   5.3229   1.8639   2.5232  -7.1708  16.2405 -11.3207
#>   -3.2248  -3.2667  -0.3067  -1.9484   8.7055   5.2577  -1.4998   0.4569
#>    1.6968  -4.6602   3.7998  -8.4993  12.9786  -0.3913 -21.7177 -21.6686
#>   -2.9383  -1.9188   2.4941   4.2077  -4.7628  -7.2720  -3.6235   8.8364
#>    3.9696  -3.0185  -9.8265  -0.8505  -8.8108  -5.6927 -15.7858  26.9669
#>   -3.8865   0.3206  -0.9331  -0.4784 -10.9885  -1.5159   1.3496 -17.8669
#>    0.3508   4.6504  11.0636 -18.4261   6.0470  -8.6634  22.1345   5.0038
#>    3.3199  -5.1318  -1.1878  -1.5742   5.2154 -13.7850   3.4688   7.3114
#>   -4.5449   6.6485  -0.3694 -10.4881  -5.5518  -2.9549  11.5984 -15.2952
#>   -0.2636   3.9459   2.1856   3.7963   0.8884   3.1580  -8.9346   0.3869
#>   -3.0938   0.5304  -8.7612  -6.6792   9.9151   3.2477   0.9332  -5.0622
#>    4.8716  -0.8094  -5.4082   0.1954  10.9452   8.7973  -2.5665  19.2408
#>   -6.9840  -7.8913   1.5115   3.1558  -0.1851   3.3493   9.8494  12.8905
#>   -2.0343  -3.9375   1.7099   6.7873   5.9394   4.5373  12.2124  -9.9828
#>    0.2709   5.2412  -0.5282   5.6714  -9.6397   4.9778  -1.4602   4.7528
#>    1.0929   1.5987   0.9800  -1.2880   1.5231  -0.4855 -12.9369  15.2107
#>    0.5867  -4.3712   3.1034   4.8287  -1.2461   6.0472   2.3134   8.0929
#>    3.3410   3.1006   4.8219   7.0904   2.0874  -1.0430  -0.7174  -3.8555
#>    2.8855   1.2734  -4.7883   5.4637  10.9526   2.4369   4.9841   7.4169
#>    1.4058  -0.8721  -0.0155 -13.3746   8.7384   5.3105  -8.3852   8.3663
#>    4.8534   4.1036  -0.8135  10.5539  18.5209 -11.9402  -9.8706  -8.8344
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]