Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   9.1556   4.1821   7.5573   4.6213 -16.0193   2.0074   9.9572  -4.5758
#>   -1.5476   4.4012   0.7032   1.5105   5.0811   2.8373   4.5659  -5.4483
#>    3.4353   9.0435  -0.8559 -18.4243   8.0786 -20.2560  -3.6145   2.9171
#>   -4.0464  -7.1854  -0.3659   9.2011  -4.3873 -12.6097  16.1488   0.0240
#>    2.4654   5.5360   1.5018  -5.0994   9.3024  -1.9080 -14.8906  -1.2917
#>    3.8280 -14.8787  17.5422 -10.0669  -3.4418  -8.6770   5.5317  19.5420
#>    2.7367   2.8312  -6.5673   5.5652 -14.5352  -4.3730  -0.6364 -21.7325
#>    5.3168  -4.5432   0.6460  -3.4376   8.7739  -4.7741   2.4209 -17.0810
#>   -6.9077   2.2159   1.1746   3.3205   0.1450  -5.6907  13.6361  -1.8147
#>   -1.1021   6.1951  -2.2553  -0.0164  10.3327   8.2402  -4.2131  -6.1404
#>    1.2458  -3.6470  -7.6172  -1.6844  17.6936  10.9374  -0.5761  -8.6999
#>    0.9661  -3.6191   2.0039  -5.8932  -7.2902   6.3301  -1.8522  12.9382
#>    0.1001   6.5468  -2.7169  -5.1886  -4.3486  -3.1810   8.0352  -0.9341
#>    2.3330  10.2530  -0.0104  -4.0348   2.7116  -6.5432   0.2556  -9.7293
#>   -5.8961   9.6964 -11.0540  -5.4106 -14.6899  -1.9617  -2.6592 -15.0869
#>   -2.3861   8.8199  11.3136  -6.7835  -5.3168   4.8975   2.5558  -1.2304
#>   -1.4337  -0.2260  -3.6689  -2.0193   6.6596   7.9267 -13.0081  -1.1840
#>    0.6846   3.7525   1.3173  -4.7834  -8.4138 -15.0017   7.9750   2.8094
#>    5.5281   2.1455  -1.4606   2.6036   3.8179 -16.4110  -4.7858  -3.2233
#>   -0.7959   0.3744   4.1807   5.7058  -9.5680   5.5859  -5.1564   4.6572
#>   -1.3152  -7.8659   4.5333   2.2587  -8.0326  -9.5263  -6.9323  12.8967
#>    4.5748  -2.9720   1.6879  -9.0233  16.6206   7.3796  12.2590  13.5882
#>    3.3614   1.9873   3.1187 -16.6296 -10.9526  11.6141   5.8648   4.5841
#>    1.1171  -0.6625  -0.0348   7.6193   9.0938   3.6556   5.2206  -2.2150
#>   -5.1527  -1.1402  -3.0599  -3.6535  -0.5114   7.2357 -17.1137   0.2885
#>    4.5137  -4.1905   9.2048  -4.3293 -11.6444   0.9436  -6.6101  -3.2168
#>   -5.1262  -3.4753   2.7399  -0.6550  -7.2675  10.9096   5.5818  -2.2183
#>    5.0456   3.9212 -17.3649  -1.5039 -11.9203 -10.6214 -10.5673  -9.3967
#>   -1.0378   2.5481  -5.6499  -4.7583  10.6078  -8.6161   1.3236  -4.7240
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]