Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.0935   0.3253  -1.9977   2.1276  -0.1416   4.4395  -0.5739   6.1524
#>    7.2085  -6.0858  -1.7093   5.7176   1.9858  -3.0169   1.6433  -5.1328
#>   -0.4592 -12.4228   9.8855  -1.9582  15.3732   0.3675  18.0942 -10.7099
#>   -0.9270   1.0443  -7.7511  16.5206   5.3605  -0.7938  -7.6978 -11.1039
#>   -5.7426  -0.7843   1.9721   9.4195  -1.2303  -2.5558  -1.7342   7.0965
#>    0.4542   1.9226   1.7567   3.0940  -0.9704  10.4253  -2.1034   5.3715
#>    1.6515   3.0630   1.5038   2.2061   3.1560  10.3958  -4.3365   1.2078
#>   -1.0963  -2.9324  -7.0560  13.5477  -7.2223 -16.6671  -1.8805  15.8982
#>    7.0842  -6.8309  19.7142  10.4759   8.9259  -6.5160   4.6963  12.9624
#>    2.1764   5.8722   8.5494   2.4027 -15.3983   3.7536   9.3052  -0.4352
#>    1.4132   5.2210  -2.5007  -9.9466   1.7079   0.7417  -2.0823   6.3581
#>   -3.7303  -4.0256  -2.9099 -12.4275  -0.2416 -16.4327   1.6645  -7.1160
#>    3.5109   8.6199   4.7595  -5.8966  -5.5387   8.7487  11.4214  10.2316
#>    3.6329   1.2154  -2.1886   2.9020  -3.1030   2.9223 -10.9927  -3.1361
#>   -0.9134  -3.4287  -8.9433  -3.5866  -5.3133  -1.1828   4.0443  -1.8923
#>   -7.2866  -0.9502   1.9593  13.8045   6.6772 -16.9748  -9.2892   1.7305
#>    3.1797  -6.8299   5.1209 -22.0436  -3.9742   0.9154   5.6184  -0.9251
#>    2.6434   0.8931   3.3879   0.9441   6.8649   4.2827  -5.1111  -5.5824
#>   -0.5080   1.6852   9.7515 -11.0793  -8.8731 -10.3701  -9.5195   8.8230
#>    3.3749  -0.9257  -8.0114  -7.6287   2.9676  -4.1840  -0.1865  10.8490
#>    1.3537   0.5068  -2.7031  -0.9482   0.3566  -2.9896  -2.3184   0.5956
#>   -1.3931   5.3596   0.9030  -0.2517  -8.2896  -9.5982  -9.7724   3.2251
#>   -1.3585   0.3176   2.4793 -10.0702  -1.6874  -4.2522  18.3130  -7.2662
#>    3.6156  -0.4469   1.9774  13.2739   2.0108   1.5304  -3.6971  -6.5367
#>    0.5814   8.0357   0.6273   1.1744 -10.5816   6.8624 -11.9930   2.2378
#>   -1.6369  -5.0005   2.8106  11.3185 -14.7707   1.1953  -5.7796  14.0753
#>   -2.8367  -3.0968   5.5873  -7.8935   0.2743   5.7185  20.2516  11.2570
#>   -2.4007  -2.9314  16.4384   4.3312  -1.5237 -17.6228 -10.1100   2.1789
#>    0.0794  -5.1006  -9.4995   5.2894   2.0942   0.1547  -6.5607   3.6631
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]