Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   7.9718   7.7289   1.7615  24.8333   3.4777   6.1575   9.9049   0.6414
#>    5.9875   1.4990  -5.8041  -9.7707  -0.3332   7.4391   3.0762  -5.1446
#>   -2.8656  10.1822 -11.6758   4.9470   9.5450   9.0015   7.8689 -14.2043
#>   -3.2723  -4.3930  -4.4642   6.9769  -0.7108  14.0728 -13.6012 -18.6754
#>   -0.7248  -1.4634  -1.0098  -3.3286  -5.0420   6.4859 -14.9684   7.2450
#>    3.6572  -8.3736   4.0941 -11.8375  10.1240   0.6288  16.9405  21.2866
#>    2.4087  -5.1216   1.1126   3.0091  13.0061 -18.1702   2.8670   1.0937
#>   -1.3459   7.9630   3.8676 -11.1358  -3.3791  15.8136  10.0103  -3.5656
#>    7.4519  -3.4537   6.3772  -3.4314  -1.2408  -6.1861   3.5343   7.1839
#>    1.3943   6.7978   4.5667   5.7776  -0.6385  -4.4003  -5.8096   3.3851
#>   10.5440  -9.2338   5.1985 -11.4010   3.1080  -8.2344  -3.7494  16.1029
#>   -6.0196  12.5248  -1.8384   1.1104   2.2027  -8.8425  -4.6384 -12.8050
#>   -0.0571   6.4225  -0.8242   1.5355   0.0120   1.3739   5.1495  14.4111
#>    2.5575  -5.9751   0.2963  -1.5561  17.3397   1.8842   6.2581   6.8608
#>    1.2188  -3.7457   4.9561  -2.5758 -25.1694   4.8033   0.5238   6.1838
#>    3.4902 -11.8498   0.2728   7.2769  -1.7625   3.5296  16.0649  -3.9263
#>    3.3031   6.2254   1.5161  -1.8854  -7.7730  -1.8991   5.0098   5.5643
#>    1.2374   2.2011   4.7019 -13.5548  18.5855   4.7754   0.2617  -0.8664
#>    5.9586   3.5664  -5.4839  -2.6515  -0.2950   9.0360  -4.9201  14.6733
#>   -3.7779 -11.0493 -11.7787  -5.4340  -5.4218  -8.8090   5.8908 -12.6927
#>   -0.0565  -1.4481 -10.8927   3.8100  -1.1673  -5.5951 -16.8337 -16.5506
#>   -2.5238   0.9751   3.2747 -13.1422   1.9371  -0.7420   8.9948  -3.1982
#>   -2.2928   6.2787   4.7247  -2.0423   4.4685  15.4845   7.4071   2.9548
#>    1.0114  -1.8731  -3.6838  -2.5177   3.0657   6.4291  -7.4270  12.1699
#>    3.4171   0.8421   3.8295  -5.8278  -2.7763   3.3311  -2.3768  15.0644
#>    4.8370  -1.9475   0.4971   0.8045  10.1380  10.1432   0.0964   2.2093
#>    4.3506   1.3995  10.0221 -13.5843   7.2646  -4.7091 -10.8550  -9.9506
#>    0.9430   0.0147   1.3229 -12.8446   2.4889 -16.1853  21.7777   5.5477
#>    5.7066  12.2345   3.4140  -4.3067  26.8060  -0.4185   6.6911   4.6060
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]