Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.0759  -3.2667  -3.6726  -3.6146  -0.8446 -11.9465 -13.2480  -1.5790
#>    0.3975  -1.2450  -6.5036   0.0190  10.6642 -11.2829  -1.2747   9.0557
#>   -0.6993  -3.0521   1.3064  -3.1416 -12.3629  -1.7182   5.0983 -12.9791
#>    0.6677  -0.9650   9.9869   3.2253   0.2826   2.6375  17.9360 -13.0811
#>    1.0986   6.0581  -9.9085  -5.5353  17.5015   4.2158  -8.6055   5.2478
#>    1.3524  -6.8643   0.8264  -1.6865   2.2044  14.0000   0.0815   5.6294
#>   -0.4596  -5.5108  -5.2178   6.6656  -2.4265  -6.7743  -1.4468   5.7187
#>   -0.3201  -2.1539   0.9406  -0.3206  -9.0602  -3.7479   5.3322  -3.7475
#>   -0.7195   5.5475   0.2906 -14.2941 -10.3774  -2.8364   5.4658  10.3086
#>    1.0841  -0.1436  -2.7277  -2.6527  14.8815 -11.4060  -2.9201  11.5450
#>   -2.5699   0.7163  -2.4689  10.1197  19.0297 -22.1998  -9.0818  18.3122
#>    3.8595  -1.7278  -4.7165   6.2495  -8.1213   4.2228  -4.6899   8.4046
#>    2.6943  -0.6113   6.9045   3.1091  -1.8658  -7.3759  13.1752   0.1586
#>    0.5644  -5.5353   1.7082  -1.0484  14.3052  -8.5054   2.3410  -8.3678
#>   -0.8436  -1.1228   4.1347 -14.6099   3.5222   9.7670 -13.6201  -2.4029
#>   -0.3781   1.6370  -4.5358  -0.3309   6.3482  -1.5515 -11.1978  12.7385
#>   -2.3546   1.8780   2.2374  -9.2340   5.9785   0.4375   6.9389  -5.3845
#>    1.8103   2.2698   3.9624   0.5547   8.0523   1.1047 -11.2401  -1.4048
#>   -0.4597  -0.4797   0.7950   0.4991   1.6633  -1.0754 -12.0461   5.2763
#>    0.1542  -3.4580  -3.9052   4.9878   0.6300 -14.8098  14.5169  11.8085
#>    2.2852  -0.7484  10.1593   4.5973  -8.7817   4.5382  -4.0239 -13.0100
#>   -6.4933  -2.2921   8.0175  -2.8289 -10.1720  10.9855  -5.2894  -8.3808
#>   -2.3902   1.5023  18.4548 -10.3783   3.5845  12.9917  -3.4189 -13.9314
#>    0.0719  10.4314  -5.8478   7.0760  -6.7528  10.4719  15.4596  11.3296
#>   -3.3912   2.8882  -4.6902  -8.5360 -13.5600  -2.9242  -4.4808   9.9590
#>    1.6890  -3.4273  -5.4863  -7.2070  10.7108  -4.0001 -11.4898   1.0628
#>   -0.5808   2.7761   2.1984   7.5121  -5.4607  18.0968  -4.4789  -7.6052
#>    1.9755  -0.5106   2.3326  -7.0511   2.8500  13.1398 -12.0309  -4.1229
#>   -1.5862  -0.5501  -5.9337  -0.7887  -4.0317 -13.7018  -7.4432   3.2733
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]