Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.6629 -10.2130  -3.7956   3.8271   0.9280   4.0572  -0.2386   6.0439
#>   -0.4198   3.2620   4.5110   1.5053 -14.2941   3.7607   3.0874  -2.5174
#>    0.0061  -2.6834   2.9305   6.9610   0.0386  -2.2842   1.5146 -18.8669
#>    0.0035  -3.5799  -0.9718   2.0866  -1.9094   6.5044  -1.7648 -18.9031
#>   -2.6732   3.2251   4.8352   8.0750  18.5582   5.6609  -4.8188  -1.2774
#>   -4.0618  -1.3984  11.3201  12.5699  14.0949  21.8155 -10.8628   1.9079
#>   -3.5575  -1.5941   4.1604  13.5323  -3.5585  14.0552   4.7872  -4.8078
#>   -0.1480   7.9138  -2.8671   2.7931  14.5851  25.1269  18.1272   1.8802
#>    2.8332   3.6290   0.4258 -11.8340  -3.7724 -14.6980   6.7259   5.7678
#>    1.2866  -4.4074 -10.3993  13.8618 -15.0700   3.5918  -5.4169  -9.2402
#>    4.0036   5.4733  14.5678  -3.7117 -11.0085  -4.5848  -4.0594  -6.4896
#>   11.4891   9.3100  -4.6211   2.4894  -3.5999   5.5650  12.1082 -12.7429
#>    5.3825   4.4967   4.8163  -3.2149  -6.3209  -1.2647  -3.7325   4.3179
#>   -5.8510   5.2655  12.4232 -12.7123  12.2282   1.1565  -3.6632  12.5140
#>   -4.0528  -4.7853   5.2860  -2.7867  -2.9223  -2.4874   5.2468  -7.9495
#>    4.3573  -2.7670  -2.4525   0.5951  -1.6829   0.1935  24.8069  -6.2803
#>    1.1327  -5.0200   0.4018   5.2509   1.4384   7.1223 -11.8343  -1.8872
#>    4.5457   9.6215 -10.6467  -0.4952  22.0520  -1.0302   0.1983   4.9781
#>    1.7578   6.2579  -2.9108   2.1739   3.9752  -0.1158   9.0949 -15.8957
#>   -0.6700  12.0175  13.2710   6.7284  -3.6497   5.2673  -0.7208  11.9823
#>    1.6242   2.2815  -0.1498  -1.4383  -6.6318 -15.2929  22.3862 -17.3294
#>    3.5452   3.1266  -7.8967   4.1302  -4.7988  -5.5148 -10.6452   4.4427
#>    3.4648  -2.5824  -6.4205  10.0700  -2.4258   8.1987  -3.8131  -7.0995
#>    0.7306   0.3054  -1.9760   9.0703  -9.2323  -4.6515  -0.4038   7.2848
#>   13.6599   9.4666  -1.3890  -4.5790   7.3976  -9.7291 -20.8730  10.3877
#>    0.9048   6.6543   2.5346   5.7222  -4.1469  -6.6989  17.7489  -1.8699
#>   -1.2639  -1.9911   0.2473  -5.9124   2.6435  -0.8537  -4.9472  19.7944
#>   -0.1906   4.2508  17.1607   0.3391 -16.5353  -3.7782   3.9446   2.1473
#>   -7.5201 -17.7539  -1.5775   8.1356  -7.2188 -13.0503 -12.5961  -9.2567
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]