Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.6727   0.1192  -4.3987   7.0252   6.8092   1.6080 -10.6994   3.9850
#>    0.2029  -4.5579  -7.0931  -7.2102   0.4172  -7.3766   9.3775  -2.6577
#>    3.0413  -0.0554  -1.1783  -0.3509   8.8694   6.0089 -19.9481 -13.4111
#>    1.3916   0.8643 -18.1407   3.3887 -10.4591  18.6929  -9.6859  17.3288
#>    0.3229   4.3975  -2.0533   6.9262   8.9837  -1.8906   3.8987  -4.9907
#>    4.3568  -4.7234   1.9044   9.7775   2.3407  18.7977  -9.4381   6.3550
#>   -4.6469  -1.4626  -8.9930  -1.6111  11.5120  21.5099  -8.9146  -6.8968
#>   -3.3905  -6.0153  -7.0299  -1.6193   1.5221  13.7266  -4.8301   1.5131
#>    1.9716   1.7007   7.9787   5.5429   6.8299 -10.6201   3.9595   5.9296
#>    1.2723   0.0104  -4.9537  -3.8492   0.7739  16.8779   4.3218   9.0329
#>    0.2579  -8.6496  -5.7006  -2.5891   9.9624  -2.8471  -4.0552   3.0731
#>    3.2242  -1.5575  -1.7073  -7.4886 -11.0264 -10.3401  -1.4712  -0.7051
#>    3.0927   5.0967   6.2714 -13.2655   4.8636   8.6445  11.4208  -1.2200
#>    2.4140   0.2434 -12.2880  -1.7525 -14.7868   9.0144   1.2601   6.2700
#>   -3.3607  -6.0400   2.2473  12.2224   7.0810  -0.8460 -15.5354   3.6009
#>    2.8606  14.3054   7.2603   8.1418   5.5081   7.9346  -5.7451  -9.0893
#>   -0.3698  -2.8294  -5.5283   3.4738  -8.1866   8.3412 -11.2000  -1.9916
#>   -3.8681   1.4643  -2.3700 -12.6426 -18.6907 -14.2242  -2.5473  -0.4450
#>    4.4514  -0.3187   1.5502  12.0794  -0.5299  18.2675   3.6964   5.5843
#>   -3.7009   3.0764  -7.9838   4.3321  -5.2129  -9.3553   6.1020 -12.9355
#>    2.0018  -4.3122  -8.3724 -12.5494  -9.5402  -0.6149  -0.5846  -7.4759
#>    0.3608   3.9122  -4.8073   5.4832   2.0321   1.9910   2.3617  14.2018
#>    1.0036   1.6773  -7.4009 -10.0153  -3.4628  -5.1606 -14.6337   8.1390
#>    2.7669  -2.5691  -4.6894  -3.9122   1.8557   9.9614  -4.0521 -19.6443
#>   -5.6178   4.6105  -7.0407  -2.8839 -19.4699 -10.5012  -0.9728 -11.4685
#>   -2.7024   3.3365 -14.4927  -8.3286   3.6445  -8.5923  -4.5629   0.6534
#>    3.8144  -6.5291  -5.6442 -17.1706   5.9722   4.9815  -0.0978  -9.7678
#>    0.0331   5.0234  -2.6833  -4.3175  -3.0290   5.3871  19.9677  -0.8033
#>    3.6551   7.4236  -5.1422  15.8916   7.1412  14.8025  19.1219   5.4485
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]