Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.8798  -1.7457  -0.9352  -5.4106  -0.1482   1.9095  -9.3355  -3.2143
#>    2.1653   2.8505   0.0799   6.5109   2.2115   5.2339  -4.7964   1.9409
#>   -5.8879   7.6103 -10.9364   7.7740   7.3133 -11.7859   1.8018  -9.9045
#>    4.7761   6.4062  -5.3729  -8.8922 -10.8351  -6.6454  -5.4254   2.9122
#>   -3.2136  -9.1434  14.7673  10.1399   8.2681  -8.3763 -10.9632  -2.0260
#>   -3.0280   4.7680   0.0077  -0.7973   3.2505  -9.8860  11.1902   5.0529
#>   -1.4242  -2.6542  -6.4440   3.0936 -11.7059  -4.5155  -1.5068  -3.7339
#>   -1.2906   5.7164   3.2404 -10.7415 -17.7528 -11.3748  -4.9849  -7.2254
#>    1.3139  -2.0989   5.2090  18.5054   6.3504  -9.7789  17.3360  -3.8080
#>    2.2641   1.6498   3.6060   7.6386  -3.3258   2.7100  -4.0967  -2.6493
#>    0.9742   0.6043   0.4904   0.6115   1.3052  17.4654   4.4823   2.9080
#>    2.0770   4.6411   0.5137  -1.2844   2.2519 -11.6329  -7.5275  -5.7659
#>    1.2895  -2.3742   6.7750  -7.5195   9.2978  -2.8145  12.3155  -8.9601
#>   -1.5170  11.9634 -10.4873   7.9755   6.6287   0.7590   2.7676  -1.7695
#>   -2.2483   7.0646  -0.1190   3.9584  -2.0451   7.0600   8.8098   3.1616
#>    7.0696   1.4495  -1.5489  -8.3575  -5.2024  -9.3643 -11.4227  -0.8878
#>   -0.7741   6.0322  -3.4977  -3.6356   0.7300  -1.4538   5.9310   6.5786
#>    8.2663   8.9483  -4.2377   0.7747   9.7970  -3.6255   0.9352   1.4329
#>    1.3675   3.0608  -1.6995  -0.1975   2.0121   0.8017  -2.5393   8.0677
#>   -3.5738  -1.6395   0.7591   6.3897  -2.0107   1.1232 -15.7523  -2.5624
#>    5.1095   3.0111  -0.6829   5.1756   9.1582   0.4750  14.0184   9.8532
#>    0.4457   8.3262   0.7579  -5.2525   9.3962  -1.9794   2.1084  22.9421
#>    0.6255   0.2309   1.4958  -3.9492  13.8083   4.1498  -6.8344  -6.6609
#>    6.7843  -2.6282   7.1541  -0.4696 -12.8962   1.6478  -3.9775  -1.1960
#>   -1.5684  -3.2593 -13.1141  -1.4237   9.3299  -5.0775   7.2893   0.3467
#>    1.0171  -3.6960   2.6941  -2.6452 -16.2423  -8.6082  -9.0763  -2.9277
#>    2.3023   2.9148  -2.3498  -1.7704  -2.4872  -8.2413   0.7641 -16.3813
#>    5.4688   8.2775  -1.2820  -8.3229   2.9766   5.6783  10.9950  -2.8085
#>   -5.4705  -7.5522  -6.8185  -4.4284   8.6162  -4.9614   4.3687   4.7290
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]