Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.3067   0.0706   4.6586  -1.6954   1.0140  -4.0335  -0.3581  -8.9362
#>   -1.4409  -5.6533   4.3828  -4.8662  -3.4792 -14.1931  12.0613   3.9767
#>   -0.0174   4.2492  -3.6086 -14.0055   3.4285 -11.6099  -5.6610  -5.7694
#>   -2.5394  -2.4514   0.3978  -6.4755  -1.8466   4.4801  -1.3162   0.4835
#>   -2.0209   3.1364   0.8799   9.9197  -0.4806   9.1804  -3.4062  10.4283
#>   -1.9160   1.2778  -0.2107   2.1122 -10.1983 -13.2995   5.5373 -12.9751
#>    3.4933   5.8347  -1.8060   3.8305  -0.1438   1.8887  -9.5208  13.8264
#>   -2.8396  -7.8725   3.0477 -14.5670   4.9887 -14.3880   3.2975  -2.1708
#>   -2.3328  -7.4518  -8.0465  -4.5519  10.4072   1.1270  -9.8714   3.5744
#>   -2.6029   8.5930   9.8088   4.1148   5.1171  -3.9222   1.9427  -6.3049
#>    3.9285  -4.8396  -5.3391  -5.4856   0.7725   3.3463 -11.5913  -4.3448
#>    0.7304  -0.8010  -5.1768  -7.5210   5.2248 -13.9490  17.3271  15.4173
#>   -1.6713   1.9790   2.3761  10.6873  -6.3259  -0.0686 -19.0850  -3.3285
#>   -4.2661   4.6511   1.9255  -7.7699   3.1793  -4.7656  -2.1572 -11.7832
#>    3.4853   2.3403   2.2622  -1.1206  17.3168  -5.5850  -9.9755  -8.4870
#>    3.0279  -2.5970  -0.4933   5.8789   8.6486   3.0904  13.8322  -4.8514
#>   -0.7414  -0.5640  -0.0589  -3.9215   7.2118  20.2233 -11.3889  -6.6035
#>   -0.6824  -5.6165  -7.8793  -4.8206  -0.5690   3.8291 -11.5274  -7.6704
#>    3.2973   6.7165  -0.1281  12.7156 -11.8339   4.2125  -7.8274 -19.4026
#>    4.5268  -0.7440   2.8278   7.3538  -1.0878  -4.0991  -7.0006 -13.5444
#>   -0.5341  -6.1985  -4.8711   1.8891  -4.6606   9.2898  -0.2814  -1.4322
#>   -0.3096   6.5935  -3.7321  -5.6055   0.3436  -6.6581   1.6337   0.0186
#>   -4.0565  -2.5248   3.2951  -9.6501  13.5765 -12.7611  -1.9917  12.7381
#>   -1.2657   1.7623   8.4412   5.6137   3.2758  -0.8042   2.5212  -8.5227
#>   -1.4563   0.3044  -1.8552 -12.0037 -16.1385  -5.6912  15.0086   4.6213
#>    1.4235  -0.2104  -8.0241  -4.7243   7.0677   0.0158  13.3064  -9.4170
#>    5.0623   8.0850 -13.9755   4.5012  -5.9213   4.6353   3.3339  -5.1597
#>   -1.9159  -0.5745  -0.1946  -2.0277   7.4183  -2.7825 -10.7122   3.0726
#>    1.3097  -3.7734   4.4802   6.3545   4.5733   3.8980   3.1289   7.0081
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]