Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.9157  10.3578  -6.8859  -1.3613  -9.5376  17.7482  -1.0152  16.2681
#>   -3.2730   1.6896  -3.5397  -5.6152  -1.6862  -3.4409  -1.4588   6.9460
#>    0.1094  -5.9862   8.3359   8.0704  -4.8960  10.6964  17.1141  14.4682
#>    4.9224  -5.5136   0.4079  -3.0619  -2.6444   6.7959   5.5537  -1.2446
#>   -1.7359  -7.0187   1.1925  -4.0375   4.1122  -2.5226   7.7255  11.3051
#>   -0.7903   0.6806  -2.7852  -5.5432  -0.9042 -22.8871 -18.5689  -0.2358
#>    2.3161  -1.7302   3.6561 -17.1368   9.9460   3.0462  13.2113   0.3942
#>    0.4198   2.0474  -9.7678 -10.6883  -2.9172   8.9509 -11.6916  -7.2649
#>   -0.4872  -6.0983  -1.9417  10.1504  -4.5471   0.5210   8.0788   6.4802
#>   -2.2971  -7.9114 -10.6199   1.9173 -10.6751 -10.0248 -10.0644   9.8896
#>   -0.5506   8.2218   5.1922  -9.5339  -4.1995   3.7878   7.9437   1.8509
#>    3.0881   5.8768   5.7032  -2.7615   0.6249   3.3976  -3.9289   2.7058
#>    3.9787   4.1893   6.0556   1.4428  -2.9909  12.3448  13.7838 -14.2541
#>   -5.9630  -5.3795  -2.6880   0.6288 -16.0983   0.7436 -13.1594  22.3088
#>    4.8546   2.4027  -6.8063  10.1892  18.6811  -5.0170  -1.4959  -0.0592
#>    1.1137   7.6200   6.1478  -2.6438   0.9405  -2.3335   7.8335  -5.9061
#>    0.4404   5.2201  -4.4793  -0.2892  -0.2471   3.0003  13.2533  10.3103
#>   -4.6310  -2.1921   5.7076   1.1403  -3.3435  -0.9508   2.3066   0.6955
#>    1.9464  -2.8404  -0.4383  -0.5969  -6.0076   0.7812   2.3334   1.2628
#>   -2.6429   3.7756  -4.1308  12.1105   2.7703  -0.5216  -2.5102   3.8208
#>   -0.6330   0.1628   1.3898  -1.8505  -0.3107   8.5606   3.2138  -2.8519
#>   -1.5230  -2.5100   1.7158   4.0494  -4.7386  18.2887   7.3544  -2.1360
#>   -2.5869 -14.5242  -2.9604  -1.4167  -2.4821  -7.6815   0.3748  -2.2988
#>   -0.8303   8.9045  -4.3993  -1.7684  -5.4840   7.2761  -5.2813   9.5340
#>    0.9125  -2.6459  -5.4350  14.8967   8.3768  -4.6267 -12.1285  -3.3940
#>    5.6207  -5.1469  11.1130  -1.3225   9.6999  -4.8858  -5.1870 -13.7646
#>    0.7177  -3.0178   7.1213  -9.5784   0.4989  -3.5327   0.2346  -1.9042
#>   -2.3967   1.3640   0.5784  -5.0469   7.9530  10.0066  -2.4765   3.1844
#>   -1.9452  -4.6391   5.4337   5.7674  -3.6265   5.6539   3.2390 -11.3010
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]