Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.4615  -4.2207  -5.7796   7.7905 -11.9840  -9.1285  15.4651   3.8241
#>    2.3409   0.6904  -4.3506  -4.9062  -1.4823   5.8396  -5.6592   7.2860
#>   -6.1945   3.5290   0.8217  -0.2173   1.7498   3.3445 -12.8678  10.8060
#>    4.9281  -0.6172   7.4167  -7.0511  -1.8670  -5.9847 -12.0293   6.0620
#>    4.8045  -0.1261  10.9152  -5.0322   0.9691  -9.7486   1.8381   3.4096
#>   -1.4625  -5.3990 -12.2305   4.6149   4.1510  11.0944  -4.4039   3.1864
#>   -2.1216   5.6414  -1.1118  14.9236   1.1256  -5.2562  -2.2428   8.4360
#>   -4.2332  -2.1787  -0.1970 -13.3084   0.0665  -0.3092   4.9394   3.9743
#>   -5.5809  -8.0981  15.5577  -2.7939  10.0153   2.9425   8.0474 -19.0659
#>    4.9288   9.5310  -1.3528   3.4166   3.4345   2.5450   4.3111 -18.4424
#>   -2.0568   5.6952  -3.7705  -3.1764  10.7695  -7.5384  -8.7663  -3.8865
#>    4.3516  -2.9396  10.4646  -1.1012  12.5249   3.0124   5.1326   2.7266
#>    1.5771  -5.0328   9.9952   6.8607   0.4023   9.4127 -11.4652  -0.8621
#>    2.8409  -2.9341   1.5933   1.2865   2.0252 -12.3845   5.8087   3.7294
#>    2.5555  11.7218  -1.2563  -3.5825  15.9397  -1.2883   5.8443   4.9676
#>    0.4708  -3.2841  -7.0380 -19.1582  -3.9601 -16.6022  -5.7878 -11.5719
#>    1.4448  -4.7377   8.3327   3.5215  -1.0575 -18.4785   1.5997  -2.5153
#>   10.9570   0.8310   2.9230   8.8415   0.6201  11.6397   2.2150   3.7324
#>   -2.3228   9.3070   1.6689  -4.3023   5.0669   8.4907  21.3255  -3.1431
#>    3.7464  -5.6867  -8.2606   7.1647   1.6916   6.0685   6.7495  -3.7299
#>   -0.3530   5.6314  16.5676  -0.5469 -21.4437  -3.2921  12.7742  -1.1477
#>    4.8243 -10.0645  -1.0904 -13.7013 -20.6179   7.4749   9.9101 -16.8219
#>    6.4659   1.4242   6.7429  -3.7501   2.7247  -7.0725   3.2128  -0.7402
#>    2.9349  -2.1409   8.8543   0.9256   8.6323  -0.0697   2.7678  10.2274
#>   -1.8960   0.8710   1.3512  14.6078   9.4193   6.3205  -4.4511   7.5670
#>    0.5856   0.0815 -12.0708  -5.0977   4.6692 -19.9177  -2.3437   7.0098
#>   -4.8163   0.0460 -19.0834  -0.6811  -3.0162  -9.5038 -18.8527  15.0208
#>    1.1772   2.7124  -6.2022   4.6527  -4.4135   8.6475  -6.0822  15.1036
#>   -0.7819   0.7884   5.3194  -7.8270   7.3277 -27.0120  11.4031   1.0773
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]