Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.1378  -1.4024  -1.8465   7.1573   8.0514   1.4498  -4.4574  -5.9858
#>    0.1627  -4.7812  -2.2718   4.9933  -7.8907   0.8609  -2.0083  -7.8783
#>    8.4179  -1.5647  -0.9347   2.1365  22.8301   3.6618  11.7038  -3.4192
#>   -2.1073   5.5523 -13.7946  -1.1004 -11.4393   5.2331   1.8835  -5.4237
#>   -5.5805 -14.7547   5.8199   6.2396  -4.5880  -2.8611   5.6833   1.1099
#>    7.7794  -2.9618   9.4717 -12.8097  -1.4823 -14.5870  10.7412 -16.9918
#>   -3.8665   9.3089  10.4060  -3.2843  -3.8854   6.1749  -2.2272  -8.8711
#>   -5.1978  -2.0525   5.9829  -4.4134   0.6965   1.2489   9.2327   2.1361
#>   -8.1259   9.1117  -1.9057   3.9804 -18.3156  12.6824   2.3224  -8.5093
#>   -3.3430  -1.4935   4.7506   8.9323   7.8044   8.2042  -7.5287  -2.6639
#>   -0.6877  -4.8971   4.5531  -0.2896  -4.5374   6.5064  -0.8828  11.6356
#>   -1.5894  -7.5116  -4.2956   3.0993   2.8298   2.5086   6.3362   6.3417
#>    7.4733   7.0668   5.1991   3.9809 -10.2540  -1.7540 -11.1500  10.2353
#>    3.2418  -4.1633  11.1913  -1.5392   5.2221  12.9139  -6.8463 -16.9385
#>    5.9751   4.3685   3.7797  -0.1952   2.3752  -7.6463   9.0462   7.6522
#>    5.9892   9.0124   6.5952  -3.3444  -7.3789   0.9285  -2.4917  -2.0193
#>   -5.9041  -0.5375   3.1107   6.2871   7.3116   9.8294  -8.3413   0.5657
#>   -4.0529  -5.8124   3.4723  -1.3153  -1.6678  -4.0587  -3.0356  -2.6138
#>   -1.0704  -0.8079   1.2729 -11.7005   3.3157   0.0180   1.8871   0.0150
#>   -3.2350  -0.1012   8.8618 -13.4263   4.2627  -1.2068 -11.4347  -5.0929
#>   -0.8093   6.0828  -0.6106   8.7036  -9.5404   3.5992   1.2524  -0.9478
#>   -0.8361   1.2491  13.3620  -5.4078   0.0406   6.3666  11.8819 -16.6044
#>   -2.0568  -4.9870  10.3286  -7.6945  -7.4689   0.9130 -13.1782   0.3023
#>    4.9359  -3.6901  -9.2847   4.3150  13.8594 -23.9700   3.1661  -8.5457
#>    4.5484  -2.1209  -7.1064  -1.0047   0.8924  -8.3054  14.1260  -1.0512
#>   -0.2980  -3.2199  -7.1752  -3.2743  -3.8763 -12.4253  23.1775  -2.5581
#>   -0.8057   1.5553  -1.3826   1.7924  -4.6967  -1.9597  17.1099  12.5220
#>   -4.7334   3.7904  -6.5496   9.5892   0.4093  -5.8913 -13.8986  20.8665
#>   -2.8296  -2.6899   3.1210   6.3862  11.8403   0.3909   7.1259  -3.7211
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]