Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -4.3478e+00 -1.6348e+00  2.1983e+00  7.9964e+00 -1.4861e+01 -4.5402e+00
#>   7.9349e-01  1.2508e+01 -8.2797e+00 -7.5898e+00 -1.3123e+01 -1.2663e+01
#>   1.1431e-01  1.7129e+00  3.7591e+00 -6.8183e+00 -1.0201e+01 -4.2070e+00
#>   1.1292e+00 -9.7484e-01 -1.8840e+00  3.3157e-01 -1.3473e+01 -2.9172e+00
#>   8.8681e-01  1.1616e+01 -1.2808e+00 -1.8282e+01 -1.7963e+00 -8.9365e+00
#>  -2.6996e+00  6.4035e+00  1.0233e+01 -5.4198e+00  9.5695e+00 -2.1136e+00
#>   2.4954e+00  6.8152e+00  5.9127e+00 -1.1958e+00 -3.6005e+00  3.8479e+00
#>  -9.3898e+00  5.6786e+00  1.9148e+01  3.7602e+00  6.8891e+00 -5.1077e+00
#>   4.2892e+00  7.9495e+00 -1.3217e+01 -6.3313e+00  2.3341e+01  7.8199e+00
#>   6.0306e+00 -4.3142e+00 -2.4263e+00  3.3742e-02 -1.5196e+00  2.6284e+01
#>  -6.3954e+00 -8.0964e+00  4.5264e+00 -6.4257e+00 -4.0831e+00  1.3089e+00
#>  -8.3622e+00  1.0730e+01  9.9684e+00 -4.0427e+00 -5.3344e+00  2.2950e+01
#>  -1.0119e+00 -1.0850e+01 -7.0993e+00  6.5358e-01  1.5854e-01  1.3349e+01
#>   4.9535e+00  3.3769e+00  4.0442e+00  3.3076e+00  5.8513e+00  1.4950e+01
#>  -3.7951e-01 -3.3700e+00  2.8471e+00  4.1281e+00 -2.4985e+00  1.6993e+01
#>  -4.1898e+00  1.6312e+00 -3.9892e+00  1.9644e+00 -7.0017e+00 -4.0673e+00
#>   1.9405e+00 -2.8935e+00 -1.8338e+01 -5.7833e+00 -1.8296e+00 -8.5469e+00
#>   3.9160e+00 -2.8400e+00 -1.2542e+01  1.1771e+01  6.3604e+00 -4.3891e+00
#>   1.5294e-01 -1.4631e+01 -1.6018e+00  1.6136e+01 -1.0392e+00  8.2209e+00
#>  -1.2391e+01 -8.6454e+00 -1.7169e+00  1.3986e+01  4.7515e+00 -1.0996e+01
#>  -3.2425e+00  4.8016e+00  2.1908e+01 -1.8972e+00 -9.0223e-01  6.4154e+00
#>   2.0433e+00 -8.0027e-01  5.1861e-01  1.4272e+00  1.0199e+00 -2.9793e+01
#>   1.8851e+00 -6.8729e+00 -4.8971e+00 -8.5747e+00 -9.4536e+00 -4.7877e+00
#>   1.8972e+00 -6.4435e+00  5.7894e+00 -4.8211e+00 -7.2754e+00 -5.5327e+00
#>   5.7601e-01 -3.4710e+00  1.1613e+01  1.6295e+01  1.0531e+01 -8.5398e+00
#>  -2.2495e+00 -6.2964e+00  5.8075e-01  7.3552e+00 -3.5327e+00 -9.3756e+00
#>  -5.5853e-01  3.3687e+00 -3.1452e+00  2.5291e+00  9.7707e-01 -1.5993e+01
#>   9.0413e-01  2.7172e+00  3.7454e-01 -1.2085e+01 -1.5735e+01 -8.5115e+00
#>  -1.0485e+00 -1.7726e+00  6.3214e+00 -1.0852e+01 -9.0372e+00  2.8365e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]