Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.4124   5.0660  -9.9401  -4.5538  -0.8782  -3.0205  -4.4927   5.1342
#>    3.9291   4.2477  -0.8199  -7.0602  -2.3611   2.5192  -7.1935  -3.0824
#>   -2.7717  -6.1916   0.2370  -4.0425  -9.6330   2.5755  -1.5796   1.9465
#>    1.5893 -11.5808   2.8055  -0.0945   9.7380   5.4134  -0.7136  11.1148
#>    4.7403  -1.8069   3.3295   4.9224  -7.7928  -7.5383  -9.8776 -12.5698
#>   -7.1468   5.4914 -15.9292  13.9233  12.5239  -0.2604  16.1915   6.0662
#>    5.7687  -4.3591  -5.6444   3.8127   6.0787 -13.4043   8.3391 -11.2203
#>   -0.5947  -0.2925  -5.6379  14.2711   0.2443  -5.6047  10.6659   1.9444
#>   -1.5210  -3.2984   2.2305 -10.6450   4.8980  -0.4658   7.2864  22.5543
#>    0.3834   8.8210  -0.0980  10.1593 -14.3387  -2.9652   5.1484   8.8185
#>    1.1597   2.0106   2.1223   3.9245  -7.9137   7.6182  18.9350   9.8914
#>    4.8017   7.1307  -5.4647   1.4852   1.4911  -2.9470 -11.0046 -16.8022
#>   -4.4137  -0.2698   3.8200  -2.9958   1.6590  16.0774  -1.9803  -3.2245
#>   -6.9053   3.4135  -7.7581   4.7586  -7.9113   1.8811   7.1362  17.5661
#>    2.2071   3.7982  -2.7250  -2.6890   7.3594   0.1401  -1.3948   1.5738
#>    0.6182   1.6967  -0.5778   3.2158 -16.1099  -0.3333 -13.8189   8.3679
#>   -2.3521   5.9323  -9.8184   4.0014   3.8476  -1.7462   9.8816   6.5726
#>   -4.3938   6.3991  -3.7862  -0.8838  -2.3089   8.3822  -0.0294  -8.0135
#>    5.9604  -9.8821  -7.6931   8.9331  21.0319  -4.9846  -6.0975  -7.5655
#>   -0.6292  -3.9824   3.4740  -6.0422  -2.3036   2.2046  10.7656  -6.3186
#>   -0.7333   4.6172   3.8819 -10.9167   7.5942  -4.7613   9.3820   5.3801
#>    0.7856  -4.6214   4.0985 -12.8177  -8.2803   3.5675  -4.4208   0.5909
#>   -2.6095   2.3299   6.2392   1.0682  -4.5243   5.3758   5.2396  -7.8205
#>    4.1053   0.8082 -13.5619  23.9534  -5.0034  13.0302  -8.0917  -9.4050
#>    3.0447  -0.8576  -4.3190  -0.4913   2.0852  -5.7944  -0.1212   2.3373
#>   -8.7374   2.6799   3.9002   3.4360 -10.7571   4.6949   1.0925  -6.6575
#>   -7.1299  -4.4121   0.1464  16.1619  -8.6163   5.8838  -2.9353   8.2082
#>    2.6690  -0.4614   4.4660   4.8678  -6.0145 -10.4429 -20.4682   9.2958
#>   -2.2822  -3.1122   0.9479   0.6788  14.2376  -1.7413   3.5599  -8.2579
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]