Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   9.4394   3.6435  -3.2443  -5.4352   4.2323  -7.8628  -6.7714   5.0325
#>    3.4409  -5.0405  -0.2166   5.2072  -4.2823  -3.0914  -1.2562  -1.3961
#>    9.3331   6.8695  13.0201   3.4428 -27.9233  21.2193  -1.8442   2.0553
#>    0.1963  -8.3149  -2.6233  -6.6344   7.3602   0.7390  -7.6373  18.1292
#>    1.4518   4.7101  12.4454  -9.2284  29.7000  -3.0777  15.9457   8.0171
#>   -1.5380   7.3493   7.1960  -1.6606 -17.6208   1.7931  -7.6827  -6.8491
#>    7.8023   6.4736  -5.3957   4.7676  -7.5110  11.2551  -1.8596  -7.3418
#>   -2.6190  10.6348   3.8889  12.1164  28.0573   1.7497  11.8786  -2.3718
#>   -4.9972  -3.9735   2.1603  -1.9324  -9.4395 -19.4574  -3.5715  -4.9242
#>   -3.4317 -14.7952   0.2194  -1.5307  -2.9689  -3.7455  -2.4949  -6.2613
#>   -0.2186   0.4452  -2.4609  -6.0752   2.5792  15.6316  -3.3349  -1.1800
#>   -1.3689   7.6753  -3.1033  -6.1709  -7.4832   7.0665  12.1811 -14.2246
#>    2.3015  -4.4915  10.3242   6.3913  -6.5260  -5.6155  13.2075   2.3907
#>    5.0518   3.6785  -1.8242   8.0718  -1.3119  -3.2555 -11.3889  -6.2919
#>    4.3066  -2.9793   4.1897  10.0596 -10.3407   5.5168  -3.9683   8.0107
#>    6.1783   0.1047  -0.8670  -1.8703   6.8122 -13.9876   2.3813   4.3698
#>   -3.8464  -1.8736  -6.0214  -7.2627 -10.9452  -6.1128   5.4980  -7.1268
#>   -2.6163  -6.0113   3.6533  11.4685   6.7303  -5.6813  -3.7573 -15.6935
#>    3.2805   0.2746   9.2578 -13.8536 -17.2180   7.0333  -8.4621  -4.3864
#>   -0.2411   6.8359  -7.4474   4.9622 -15.4639   2.5949  -3.8835   7.0785
#>    4.6515  -2.1167 -12.8759   0.0066  21.4958   7.2881  -9.1760  -2.7012
#>    8.8867   0.1493 -15.6700 -13.9844   3.2177   5.0364  -3.4612   7.1330
#>   -4.1990  -0.2793   6.5852   9.3839 -10.0640  -8.9824  -2.8600  -0.4223
#>    9.1881  -2.6854   0.8319 -15.3998 -10.3204 -11.0228   1.9607   5.3852
#>    1.5064  16.3627  14.0056  14.0492  -0.4169   2.9148  -4.7745  20.8929
#>    4.1464  -1.8862  -8.9185   0.0941   6.2005 -10.0723 -12.3684 -10.9533
#>   10.9129  -0.3214  -1.9015  -9.4217   9.4202   2.7848  -0.0222  -1.3093
#>   -2.0443   0.1974   0.0474  -7.1888  -1.7296   0.6017  -7.2566  -3.0351
#>   -8.0660 -15.0294  -7.1956  -9.0426   1.6682 -18.1867  -1.3196   5.9862
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]