Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.4555  -0.0106   2.8268  -9.2702  11.6841   6.8857   0.5470   2.5833
#>   -8.1520   2.3002  13.6597  -5.7456   1.0810 -12.0516   7.8770   4.6657
#>   -8.0303   8.3082   5.7355 -10.2379 -16.3520  -5.6311  14.7954   9.6191
#>   -6.4450   0.3423   3.3623   0.1703   4.1388  -5.7258   4.0197   2.6087
#>   -7.6473   4.4833  13.3075 -16.7584  -6.7230   3.3574   1.6048   9.2692
#>   -8.0812   7.1466  -1.0374   4.3723   5.6094  11.6398   9.1465   1.0384
#>    0.2730   5.4434  -5.0839  -1.8926  -0.4705   3.0246 -11.7505  -8.7319
#>   -1.7628  -2.0886  14.0615 -13.2588  -2.4863  -7.7394 -13.6613 -10.8016
#>   -1.9324   1.7683  -4.5692   9.8493   7.1397 -10.2840  -9.6707   8.4987
#>   -0.9881  -8.5306   1.0089 -13.1484  19.1924   2.6273  -9.0058  -5.9639
#>   -0.5389   3.6524   7.9048  -2.3376 -12.5256  -3.9676  -2.0021   5.5310
#>   -1.3114  -9.4225  -3.3901 -11.9439  11.7279  13.8322  -9.4471   4.3594
#>   -8.2054   4.7454  -1.9178   3.8911 -15.7669   4.3880   3.5793   2.8688
#>    7.6948  -9.1564 -10.0909   3.3223  -9.3438  13.2877  -9.5192  -7.1223
#>   -2.5707   5.8087  -1.9545  -8.6880 -12.7150   0.1389  -5.5943  -4.6186
#>    3.8668  -4.2110 -13.4643   0.5355  -8.9101  -8.0658 -25.5846  -5.7857
#>    5.1994 -11.8392   9.1748  -1.2784 -11.2625   2.0279  -4.5395  10.4094
#>   -5.2507  13.8595  -7.0961  -3.5644   0.9573  15.6554  22.3781 -21.5684
#>    2.3582  -4.0306  -0.9800 -11.8855   0.2509  -1.1937 -13.0245   0.8691
#>    9.5101 -12.7235   4.0366   4.1599 -20.8568 -14.7200   7.4479  14.0879
#>    7.2451 -15.2282  -1.3015   0.9591  20.7144  -8.0403  -9.2278   3.9012
#>   -5.3366   0.8922   0.6256  -2.6088   6.6779  -3.1337   5.3317  -5.8878
#>    1.0632   0.8525  -4.0588  -5.7957  13.2711  14.7374 -11.1681  -5.7541
#>   -1.6228   3.7055   3.9788  -0.4012   6.1663   4.0208   4.4171  10.4128
#>   -2.4865 -14.1402  17.3692  10.8713  -2.9319   5.4324  -5.5428   6.0884
#>    1.5561  -0.3019   2.7328   6.8461  -2.2435  -4.7836  -1.6128   6.2298
#>    0.6111  -1.6849  -2.7225   5.9039   1.9995   6.5856   5.4411  -3.3605
#>   -3.0380  10.0761  -8.5855  10.4303  -7.4964  -1.6714  -4.7256  -8.5738
#>   -7.1865  10.8700  -8.8352  18.7590  -5.4184 -17.6286 -10.6273  -6.2961
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]