Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.0592   3.7349   2.6319  -4.5878 -10.5578   4.8710  -0.0549 -14.4479
#>   -2.4431  -6.7887   9.3504   4.9154   2.7927   5.1763   1.8688   7.3001
#>   -1.0631  -2.6507  -6.1674 -11.1364  -7.3589 -10.2711  -4.6715   0.6253
#>    1.3593  12.8702  16.3902  12.1470  -8.0660  -3.0959   3.3689  11.2004
#>   -2.3020  -2.5629 -10.9468 -12.0030  -3.6080  14.1747 -10.7912   0.3965
#>   -0.0326   3.3703  10.1299   9.8138   8.5760  -1.0562   7.0989   4.4714
#>   -0.1515  -7.9858 -10.6600   7.0945  12.6644   8.9028   5.3230 -11.8532
#>   -1.1176  -8.5479 -11.3129   7.7679  10.6256   0.6985 -11.5498  10.3998
#>    4.5109   1.6111  -2.8152  -0.4412   9.7832 -11.9551   0.1916  -2.7960
#>    0.2051   0.7060  -7.2936  -3.9004  -4.7239   6.8044   4.2618  -8.7809
#>    7.8350 -13.5669   0.4811  -5.0729  13.6673  -7.8701   1.0559   3.5158
#>   -0.8084  -1.3837   2.9435  12.5396  -0.1808  -3.6410   0.3000  12.1387
#>    0.7434  -5.2226   6.1629  -5.4915 -15.2301   4.7848  23.1953  10.2599
#>    3.5432   1.6478  16.0679  18.2780  11.2589   1.0624  -5.5953  -1.4705
#>   -1.8616   2.5849  -7.3029 -17.6707   0.6519 -13.9479 -13.5631   4.2600
#>    0.3422   1.2897   6.3630   3.2750 -11.3788  -0.9886   3.7510  -7.7294
#>   -0.0205   2.0669  10.9707   3.5014  -6.3944  -6.0190   3.3323 -15.7803
#>   -2.1193  -3.4202   9.1850  -6.4793   2.8043  -1.8968   7.7692   4.8189
#>   -0.1002  -8.2721   1.1784  10.2939  10.2680  21.6864  -7.9434  -7.7161
#>    1.6491  -4.2302  -1.4461  -3.0803  -8.6517 -10.4836 -29.3165  -2.5168
#>   -3.3266  -0.1425 -11.1643  -1.7620   7.7625  -5.0776  -8.1292 -18.9435
#>    0.9278   5.7613  -1.4565 -15.0270   7.9672  11.1434  -3.8210  -5.6863
#>   -2.4484 -11.7023   0.7679  -6.0862 -11.0203  -0.3740  15.8101   4.6877
#>    3.2395  -7.9292  -1.6552  -5.2128 -10.2305  -4.5960   7.7904  -8.4354
#>   -5.4407  -4.7094  -3.8125  12.3294 -11.2534   3.2243  -0.3078  -5.2669
#>    6.2776   5.3206  15.6759  11.1083  -8.6330  20.5395   2.4437 -12.3815
#>   -0.4898   1.6951  -2.4304   9.6897  -7.8225  11.5668  -5.1704  -3.9439
#>    1.3198 -10.6093   1.8976   8.9340   5.7419  25.0784 -16.9737   2.0434
#>   -2.8351  11.5854   9.9908   3.8221   9.9137  -2.0461  -1.5999  10.6956
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]