Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.5798   1.5417   8.3619  -0.5237   9.0673  -7.2272  -6.9601  -1.5803
#>   -0.2403   8.7234  15.0488   3.1472 -19.4922  -0.8381  11.5809 -18.1424
#>   -6.4274  -3.5717   1.6075   7.2617  10.1074  -5.8649  -1.1969  -6.6623
#>   -2.5212  -4.3352  -9.0128  -1.5473   5.4943  16.2704 -16.4967  -9.8102
#>   -2.6788   2.9750   3.4595  30.0454  -2.9038 -11.0985 -13.7320  -7.1836
#>   -2.8395  -4.7553  -1.1610  13.7001  -5.3923 -10.8139   3.5885 -14.0028
#>    2.9470   4.0828   6.4471   2.5735   5.5496  -8.4403  -8.0354  -1.6540
#>    1.2263  -9.7735   1.0819  -3.1898  -4.0378 -10.1755   6.4451  12.3796
#>   -0.3152  -2.3583  -0.1454  -4.0739   2.0746   4.0148  -1.6473  11.2266
#>   -1.2608   7.4255  -0.2445   3.4038 -15.1428  -9.7849  -8.8651 -15.2899
#>   -0.6684  -4.8127  -3.3137  -0.4323  -6.1556  -4.0705   1.5250  -3.8439
#>   -3.9999   1.1055  -5.8181   6.8453  -0.8528   9.9751   2.6110   1.1275
#>   -2.3650  -2.2094  -5.5500   8.5876  23.8501   9.6961   2.4755   5.3295
#>   -3.2802   9.6352  -0.1618   5.8677   9.7619   1.6205  -7.0208  -3.6132
#>    3.0117 -13.7229  -6.5291  -3.1086  -5.5357  -3.2431   6.4845  13.9355
#>    0.4138   5.5702   3.8391   8.7735  -1.0350 -16.1937   4.9366   7.1822
#>    4.9083  -4.7014  -9.0244  -4.2895  -7.1845  -4.1988   0.3462   1.9083
#>    4.0230  -6.2123  -5.1778  -1.8140  10.3883   8.8888  -8.8224  -8.9839
#>   -5.7402   5.3483   1.6061   7.1222  -6.7688  21.9224  -4.9063  -7.4452
#>   -5.3787   3.7766  -5.0120  -1.5869  11.0816  -0.7584   2.7802  -3.3936
#>   -6.0269  -0.3546  -8.8297 -11.6645  -6.1655   4.5813  36.4733  -3.1911
#>   -1.6774   6.2868  -6.6083  -0.9308   6.1691  10.0196   1.4146  -9.0681
#>    3.1378   2.1745   0.9858  -9.6086 -17.8970  -5.4960  12.7172   6.9359
#>    2.3772   5.1633   2.6407  12.7433  10.5661   7.8068  -3.9950   2.4744
#>   -0.1295  -5.7609   3.9345  -0.3616   2.8499   1.3539  -4.2077   3.2834
#>   -2.7440  -0.9769 -12.5236   9.0289  -7.5987  14.6902 -12.9580 -10.5486
#>    0.1709   1.3329  -9.4504   3.4206 -17.0332  -6.5816 -11.2215  -6.4326
#>    3.5828  -0.8957  -2.6628 -13.7812   1.5123  22.0337   4.2319   3.4897
#>   -3.5397  -2.6421  -6.2231 -20.0646   4.1220  -1.7685   4.8867  -3.6084
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]