Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  4.5121e+00 -6.7326e+00  3.2542e+00 -2.2684e+00 -5.9225e+00 -8.9014e+00
#>   7.8748e+00  4.9430e+00  7.2493e+00 -6.4845e+00  7.4703e+00  2.8833e-01
#>   4.7930e+00  5.7953e+00  6.7724e+00 -3.2300e+00  1.5753e+00 -1.0019e+01
#>  -4.7229e+00 -3.6046e-02 -3.8486e+00  6.8967e+00 -3.5864e+00  5.5427e+00
#>   4.5485e+00  1.8641e+00 -1.0395e-02  4.9277e+00  1.5645e+00  1.1354e+00
#>   1.5155e+00 -2.3630e-02 -7.8709e+00  2.6854e+00 -1.6794e+01 -3.3097e+00
#>   3.0427e+00  1.1837e+00 -6.2294e+00 -1.5539e+00 -1.8669e+00 -3.1218e+00
#>  -4.3838e+00  8.6978e-01 -5.2551e+00  1.1061e+00  1.9710e+01  9.1268e+00
#>  -4.2929e+00  4.2149e+00 -1.8673e+00  1.4140e+01 -6.5774e+00  8.3708e-01
#>   3.0045e+00 -2.1389e+00  1.8295e-01 -2.8640e+00 -1.1696e+01 -9.4533e+00
#>  -3.7156e+00  9.7469e-01 -1.5673e+00  4.3363e+00  5.2654e+00 -6.2304e+00
#>  -1.8827e-01 -5.8934e+00  4.6373e+00  4.0804e+00  1.3592e+01  8.9160e+00
#>   3.2127e+00  9.2743e+00  6.9908e-01 -1.4984e+01 -4.2706e+00  3.8572e+00
#>  -3.3050e+00  4.1082e+00  8.6693e+00 -1.7819e+01  6.0430e+00 -8.4769e+00
#>  -8.5382e+00  2.0196e+00  2.0721e+00 -6.8837e-01  3.4007e+00  6.8832e-01
#>   3.7898e+00  2.1840e+00  4.9404e-01  4.1792e+00 -2.5343e+00  1.4001e+00
#>  -3.4247e+00 -4.5094e+00 -2.3402e+00 -7.4171e-02  1.0485e+01 -2.5406e+00
#>  -1.2961e+00  1.8085e+00 -4.4724e+00  1.9026e+00 -2.2493e+00  9.2286e-01
#>   1.6272e+00  7.5802e-01 -5.7794e+00  1.2972e+00 -7.2367e+00 -1.2148e+01
#>   1.1983e+01 -3.6386e+00  7.4671e+00 -4.0164e+00 -5.9204e-01 -1.5448e-01
#>   1.9575e+00 -6.0358e+00 -5.7396e+00 -1.0718e+01 -7.2139e+00  8.9548e-01
#>  -6.1769e-02 -2.7827e+00 -3.1550e+00 -3.2635e+00 -1.7399e+00 -6.0660e-03
#>  -2.8345e+00 -3.4347e+00  7.0885e+00  1.9342e+00  8.2374e+00 -8.8482e+00
#>   3.9941e+00 -3.9584e+00  4.4221e+00  9.0901e+00  9.8718e+00 -1.1891e+01
#>   2.4721e+00 -3.6624e+00 -4.1575e+00 -1.3240e+01  1.3275e+00 -1.9107e+00
#>   9.1811e+00  1.0232e+00 -6.1434e+00 -1.2990e+01  8.4135e+00  9.9075e+00
#>  -3.0195e+00 -9.9961e+00  1.2603e+01 -2.7098e+00  6.7421e-01 -5.9052e+00
#>  -3.7589e+00  5.6579e+00  2.3266e+00  1.3200e+01 -1.2510e+01 -2.6147e+00
#>  -5.3083e+00 -1.0115e+01 -2.7405e+00  9.1653e+00 -4.8381e+00  2.9590e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]