Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  2.5673e+00  6.6661e+00  3.6602e+00  5.1265e+00  2.3030e+00  3.5304e+00
#>   1.4432e+00  4.0713e+00 -2.9548e+00  7.0665e+00 -1.6998e+01  5.3032e+00
#>   7.2590e+00  6.8392e+00  6.8125e+00 -1.1222e+00  1.4693e+01 -2.5216e+00
#>   6.0749e-01 -1.0697e+01  3.3995e+00  9.8611e-01  1.1927e+01  1.1525e+01
#>   4.8360e+00  5.4539e-01 -2.0602e-01  9.4460e+00  3.1837e+00  2.2194e+00
#>   1.6948e+00  6.5060e+00 -2.4448e+01  1.4214e-01  7.7319e+00  9.8829e+00
#>   3.4992e+00  6.5452e+00 -4.3350e-02 -1.4395e+01  2.9630e+00 -2.0436e+01
#>  -1.3472e+00  6.5177e+00 -1.0635e+00 -1.4189e+01 -8.7951e+00  3.2202e+00
#>  -1.4852e+00  3.1920e+00  1.1070e+00  9.6549e+00  6.1921e+00 -1.2495e+01
#>   7.8938e+00  1.5910e+01  4.5694e+00 -4.4206e+00 -9.1257e-01  1.4620e+00
#>   1.0819e+00  1.0879e+01  6.8531e+00 -4.4976e-01  1.7275e+00 -1.6235e+01
#>   3.3412e+00 -2.6129e+00 -3.9122e-01 -4.8633e+00 -2.1748e+00 -4.0002e-01
#>   1.2258e+01  1.6621e+01  8.3133e+00 -4.7067e+00 -7.3724e+00 -9.9783e+00
#>  -3.7803e+00  3.4487e+00 -4.5086e+00 -4.8967e+00 -5.9308e+00 -5.1883e+00
#>   1.0251e+00 -4.6638e+00 -8.6807e+00 -9.9161e+00 -1.3703e+01  2.2750e+00
#>   3.4333e+00 -9.1899e-01 -9.2512e+00  1.2847e+00  6.5626e+00  6.4654e+00
#>   1.8141e+00 -1.0881e+01 -3.6034e+00  1.9238e+00 -1.1507e+01 -7.6365e+00
#>  -3.2417e+00  8.8534e-01 -4.3251e+00 -3.9348e+00  5.0571e+00  1.2333e+01
#>   2.2403e+00  1.7594e-01  9.9148e-01 -8.1901e+00 -5.1503e+00  8.4561e+00
#>   3.8935e+00  6.8526e+00  8.5737e+00  6.7784e+00 -5.5036e+00 -2.3082e+00
#>   2.0399e+00  4.3066e+00 -2.4497e+00 -6.4285e+00 -2.1110e+01  3.5100e+00
#>   4.6825e+00 -4.4618e+00 -9.2407e+00 -1.3003e+01  2.1344e+01 -6.6196e+00
#>  -1.5023e+00  4.1668e+00  4.8371e+00 -1.2374e+01 -1.9480e+01 -3.8944e+00
#>   6.7904e-01 -4.6384e+00 -1.3042e+01 -8.8803e+00  1.0516e+01  2.5412e+01
#>  -4.4196e+00  1.7797e-01  9.3048e+00 -1.2160e+00  3.9148e+00  3.6257e+00
#>   4.8180e-01  5.4472e+00 -5.8861e+00 -2.9448e+00  4.9493e+00 -6.9413e-01
#>   3.1071e+00  1.3093e+00 -3.4637e+00 -7.2781e-01  2.8547e+00  1.8922e+00
#>  -5.1542e+00 -5.8778e+00 -3.9213e+00  5.3041e+00 -6.8236e+00  1.1808e+01
#>  -8.2074e-01  3.1450e+00  5.3353e+00  7.4855e+00  2.2807e+01 -1.0128e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]