Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.4138   5.1466   1.7670   1.2875  -1.3266  -4.1700  -7.7673  -4.2195
#>    3.7138   0.5411   3.3192   1.2000   3.3736 -17.7662   3.3307  -7.8219
#>   -3.4960  -4.1729   3.2145  -2.4075  -1.6141  13.4302   1.4110   6.8178
#>   -3.3846  -2.6407   1.9384   7.2226  -6.7073   5.3755   9.0768   7.2395
#>    4.5940  -6.8151  -2.6163  -0.2666   4.1204   8.1999  -5.8827   1.9420
#>   -3.3067  -3.6058   3.5091  -4.9926  -3.2810   0.8715 -12.7105   0.4521
#>    2.4303  -3.6155   4.2188  -2.8642   4.9873  -3.7092  13.2857   3.5877
#>    4.1827   3.3227   4.7455  10.1061   7.8300  -6.3602  10.6655   5.6395
#>   -3.0547   0.6231  -4.5132   5.5916  -1.1954 -13.1770 -19.8397  -1.1102
#>    5.2716  -3.1000  -3.9907 -17.0957   8.1347   0.6057  24.8240  11.4818
#>    2.4666  -1.2651  -2.4496   2.0002  12.1700  -5.7351   9.2832  -4.8951
#>    5.9393  -1.0744 -11.5649   9.9334  -1.1954 -14.1974   7.9559   8.9662
#>   -0.1811  10.6623 -20.8011   7.2074  21.5757  19.6854   5.7775   4.5078
#>   -0.3699  -3.5216   9.3373   5.1597   3.2456  -3.8473   5.7486  -0.8578
#>   -2.9389  -8.3945  -4.5675 -21.1329  -4.2278   1.6992  11.5701  17.0741
#>   -0.7403   5.2153   1.1915  16.7036   4.1780 -13.7508   2.2189  -9.2531
#>    0.7688   1.3705  -1.6710   8.7173  -4.5267   5.4001  -7.3500  17.6007
#>   -4.5787   0.8391   3.0991   7.1527   1.7971 -11.9086  -0.0540   8.7458
#>   -2.2755  -1.6973   3.1727   3.1396  11.9527  -8.7989 -11.9710  -1.4550
#>   -2.0815  -1.1199  -4.3233 -12.6548  -2.4463  -6.7343  -1.3802 -15.7235
#>   -4.3590   5.5850  -2.9200  -4.9862  -2.9183   1.5073  -5.8524  -3.8121
#>   -2.2759   4.5891  -1.8861  -3.3941   2.3343   6.6446 -15.0713 -13.5888
#>   -1.0865   4.3640  -0.7617  10.1450 -11.2269  21.9073  -8.4112  16.7328
#>    0.4464  -0.5035  -0.9506  -7.2539  24.2890   2.0866   0.4250  14.3716
#>   -5.9827   3.3249   4.7880   6.7125   3.0913   0.2107   0.4399  -4.4033
#>   -4.1146  -0.4527  -3.2999   1.5565   7.5408  10.3666  13.2477   0.2282
#>   -2.2372  -8.4987   0.7726   7.3973   3.0918 -13.1447  -5.4987 -12.2576
#>   -4.4115  -4.2748  -6.1667  -9.7071  -1.3615  -5.3906   9.5654  -1.8271
#>   -1.4801  12.6078   6.3076  -2.9069   0.2132   2.2714  -0.7773   5.5328
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]