Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  1.8811e+00 -6.2044e+00 -2.3053e+00  1.0531e+01  3.0012e+00 -1.6991e+01
#>  -3.1928e+00 -1.5335e-01  6.6582e+00  1.9998e+01  7.1140e+00  3.6529e+00
#>  -7.6545e-01  2.4390e+00  4.9490e+00 -7.3657e+00 -5.7842e+00  1.6026e+01
#>  -3.1531e+00  4.3056e+00 -1.5223e+00 -1.2821e+00 -3.7130e+00  3.1862e+00
#>  -1.0175e+00  6.0081e-01  6.0383e+00  7.9633e+00 -7.7544e+00  7.9035e-01
#>  -9.0100e-01 -2.6070e+00 -2.4717e+00  1.8858e+00  1.7921e+01 -5.6383e+00
#>  -1.0239e+00 -2.9110e-01  1.5115e-01  1.0520e+01  4.5285e+00  1.0091e+01
#>  -4.0288e+00  1.3679e+01 -8.5110e+00 -2.3815e+00  9.0589e+00  1.3602e+01
#>   1.0070e+00  5.3445e-02 -1.5563e+00  1.3462e+01  6.5018e+00  3.8160e+00
#>  -1.5318e+00  5.6993e+00 -5.1850e+00 -1.0859e+01 -1.3540e+01 -4.3550e+00
#>   7.8957e-01  1.0584e+00 -5.3018e+00  1.5931e+00 -5.8253e+00 -4.3425e+00
#>   6.0039e-01  1.0893e-01 -3.4015e+00  2.9698e+00  5.7654e+00 -2.6130e+00
#>   1.7641e+00  1.5034e+00 -2.2339e+00 -3.9006e+00 -8.0773e+00  2.3392e+01
#>  -2.5254e+00 -6.3432e+00  3.7986e+00 -5.2020e+00  4.6022e+00  2.9921e-01
#>   2.9692e-01  2.8223e+00  1.4710e+00  4.7040e-01  6.4827e+00 -1.9254e+00
#>  -2.2360e+00 -4.3051e+00  4.7927e+00 -1.6395e+01  9.6671e+00 -3.5654e+00
#>   1.9942e+00  4.8226e+00 -1.2470e+00 -2.1681e+00 -4.8579e+00  4.0517e+00
#>  -2.2541e+00 -1.0556e+00 -6.3678e+00 -1.9691e+00  6.5128e+00 -1.3639e+01
#>   3.0789e+00  9.0000e+00 -1.4093e+00 -1.8984e+00  9.9387e+00  4.3139e+00
#>   1.6466e+00  6.8412e+00 -1.0759e-01 -8.3774e+00  1.0302e+01  6.2382e+00
#>  -1.4721e+00 -8.0203e-02  2.4248e+00 -8.5811e+00  2.5508e+00 -5.1780e+00
#>  -1.1737e+00 -4.4986e+00  4.9307e+00 -1.3112e+01 -2.6881e+00  7.4340e+00
#>  -9.5006e-01 -5.8211e-01 -8.4074e+00  1.6240e+00 -3.9678e+00  1.8468e+01
#>  -3.1992e+00 -5.1587e+00 -3.8105e+00  6.2234e+00 -8.4090e+00 -1.6670e+01
#>   7.4898e-01  7.2684e+00 -2.7107e+00  1.6163e-01 -2.8745e+00  1.8440e+01
#>  -2.5044e+00 -4.6993e+00  5.3700e+00 -2.3158e+00 -4.2549e+00 -8.1965e+00
#>  -3.9261e+00 -3.8773e-01 -2.1442e+00  3.4581e+00 -5.3984e+00  1.8452e+01
#>  -5.0563e+00  9.2684e+00  2.2611e+00  1.0230e+00 -5.0613e+00 -1.2653e+00
#>   4.4211e+00  8.6698e-01  7.1919e+00  4.1850e+00  6.9399e+00 -4.8790e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]