Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.2463  -2.2868  -1.2213   7.0564  -7.3589   7.7435  10.2365 -20.9804
#>   -3.7804   7.0377  -1.9412   3.6729  -4.7893 -13.6279  -5.6209  10.7359
#>   -6.7169   5.8391   4.9693  10.4389  14.1028   0.8376  -3.4166  -4.5226
#>   -0.5372  -5.3883  -1.4996  -6.6876  -9.2248   0.5237  -1.6803  -7.2765
#>    4.6986   2.0875  -5.7455 -10.5321  -4.7188  -4.4992  13.1428   7.4227
#>   -4.6919  -0.7665   1.7058  17.1492 -15.9406  -0.1000 -11.8194   8.4830
#>    1.4565   1.1169 -15.0374  -0.6046 -23.9397  18.5585  -4.5533  -3.3580
#>    2.4490   3.0653 -11.9686 -15.4734  -6.1650  -5.6115  12.7453  -8.9973
#>    1.3999   3.0941  -7.9278  -0.7842  -6.3770  -1.2260  -9.7396  18.1339
#>   -6.7076   2.9804 -11.9084  15.2744  -7.6980   1.6554   2.5405  -1.2053
#>    2.4238  -4.1928   4.3552   0.9796  -7.1480   2.5755  11.4621  -8.3538
#>    2.4645  -0.2007   4.0127 -10.1352  -5.2698   1.7894   8.0595   8.3520
#>   -1.8493  -5.8538   6.5309  -3.2120   1.9954  12.6261  -6.9166   3.7972
#>   -7.2539   1.7392   7.6759  15.9216  -9.7057   2.5007   1.3821  -1.9471
#>   -4.6061   1.9051  -0.8299 -12.7243  13.4828   5.8140   1.3263   4.4070
#>    0.7830  -3.3632  -2.9026  -6.7707 -13.8546   7.3438   9.9803   1.8590
#>    1.0414   3.0225  -2.8604   6.2169  -0.6249  12.7467  -1.8809   9.2851
#>    1.9450  -5.4802   1.7335   7.5114   9.9759   4.8980  -0.9013   9.9619
#>    5.2004 -11.2131  16.0524   1.1703  -1.8681  -2.6084  -7.1103   8.5187
#>   -1.9090  -8.5148   6.2225  -0.2312  -3.4355   0.9179  -6.1440   1.2277
#>   -5.4921   9.7473  -3.0785  -5.2551  -9.3043   4.2068  10.0438 -12.5174
#>    0.8525  -2.7933  -1.3680  11.7140   2.0209 -19.7010 -12.5802   1.8968
#>   -1.0771   1.9439  -8.1939  -6.8901  -5.2790  13.7648 -12.3870   6.5320
#>   -6.1629   1.9564  -6.9233   8.9951  17.6831 -20.8376   5.6791   8.9932
#>   -3.4382   7.0494  17.2647  -3.8016   0.9471  -4.4327  -7.6870  -7.0745
#>   -4.3232   5.2215   6.5557  10.4439 -22.9132  11.5769   9.7522   0.8086
#>   -1.5669  -3.5413  -0.1324  -0.0069   9.1604  -8.2322  11.2107   1.3446
#>   -4.3558   0.0523   1.1629   1.0254 -11.8532  -2.2131   2.0990  -5.1258
#>   -3.1997  -6.8095  18.3082 -10.6340   2.3774  -3.1217   0.7062  21.5810
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]