Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -1.5239e+00 -1.6867e+01 -1.4034e+00 -9.7054e-01  1.1900e+01  1.4101e+01
#>   2.9345e+00  9.1507e-02  1.3228e+00 -3.7361e+00 -6.0410e+00 -1.0964e+01
#>  -3.3713e+00  3.5868e+00  5.3417e+00 -2.1925e+00  2.2217e+00 -4.3132e+00
#>  -2.6557e+00 -1.3410e+00 -1.1025e+01 -3.7594e-01  1.1003e+00  7.6165e+00
#>   6.9345e-01 -1.8039e+00  4.2114e+00  1.2467e+00 -6.2503e+00 -1.6726e+00
#>  -3.4161e-01  9.3396e-01 -9.4265e+00 -6.5548e+00 -9.8099e-01  4.5907e+00
#>  -2.6838e+00 -4.5751e-01 -7.2551e+00 -6.1651e+00 -5.2510e+00 -2.3162e+00
#>   1.1342e+00 -1.4868e+00  1.9004e+00  4.7617e+00 -2.5315e+01  2.9820e+00
#>  -2.0754e-01 -9.5779e+00 -7.2821e-01 -1.5253e+01  1.3300e-01  1.2094e+01
#>   3.1920e+00 -3.3507e+00  2.9064e+00 -1.3575e+00 -8.5407e+00 -2.7470e+00
#>  -3.7768e+00 -6.1667e-01 -2.5430e+00 -1.1034e+01 -8.7903e+00 -2.3618e+01
#>   8.0614e-01  1.5253e-01 -8.7855e+00 -2.3511e+00 -6.0005e+00  1.3744e+00
#>  -4.2951e+00 -1.5516e+00  3.0780e+00  1.9534e+01  2.3174e+00  1.5079e+00
#>   1.3447e+00 -1.7154e+00  6.4001e+00  8.2597e+00 -1.1497e+01  1.2124e+01
#>  -7.3862e-01 -9.7014e+00  7.0545e-01 -8.3628e+00 -6.8289e+00 -1.7535e+01
#>  -1.6931e+00 -2.0889e-01  1.0347e+01 -2.7525e-01  1.3310e+01 -8.3963e+00
#>   2.5676e+00 -3.9950e+00  5.1967e+00 -3.1056e+00  4.8198e+00 -8.5205e+00
#>   1.6143e+00 -2.0325e+00  8.8799e+00 -8.8174e+00 -1.0653e+01 -2.9288e+00
#>   3.5992e+00 -8.7614e-01 -2.0239e+00  6.8328e+00  6.5285e+00  9.0891e+00
#>   7.5804e-01 -3.7697e+00  2.0445e+00 -3.0428e+00  8.9496e+00 -2.5871e-02
#>   3.4744e+00  7.6627e+00  8.6556e+00 -2.7212e+00  1.2603e+01  1.3570e+01
#>   6.4856e+00 -2.2351e+00  6.1405e+00  1.1833e+00  9.6793e+00 -9.4925e+00
#>   5.4892e+00  2.1564e+00  5.4055e-01  2.1982e-01  6.6471e+00 -1.2633e+01
#>  -9.6525e-01  8.4210e+00 -2.5523e+00 -1.1110e+01  7.4929e+00 -1.3632e+01
#>   1.2940e+00  2.5730e+00  4.1943e-01  2.2415e+00  1.3298e+01 -3.9573e+00
#>   5.5591e+00  1.7548e+00  1.4220e+00  1.4948e+00  1.2712e+00 -2.7442e+00
#>  -3.6520e+00 -9.0733e+00 -9.5108e+00 -1.1659e+01  4.4775e+00  1.0011e+00
#>  -5.7685e+00  6.1830e-01 -6.3266e+00  1.9032e+00  8.5596e+00 -4.3991e-01
#>   3.9507e+00 -2.1484e+00  5.7397e+00  9.6712e+00 -1.2624e+00 -8.6059e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]