Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.0085  -9.2876 -10.1415   4.7217   2.7077   4.6082   0.6060  -1.0686
#>   -0.2440  -6.7657   1.2869  -3.0315  19.0525  16.1038  10.1085  13.8172
#>    1.1625  -1.7784  -0.0246 -11.3164  -0.7659   0.2996 -16.3047   2.9514
#>   -1.0630   0.0700   0.1644  -2.7793   3.6273  13.3338   7.3041   0.6338
#>   -2.7194  -0.2188  -0.2623  -3.7699  -1.7477  -0.7653  15.5515  -3.4864
#>    1.7903  -3.5863   1.4032   3.4616  10.4682   7.2838   4.4447 -17.0648
#>   -2.4013  -1.0302   2.7877   5.0019   9.4829   6.7710  12.7639  -5.7604
#>   -3.2957  -0.9719   2.6787   2.5656 -12.2298   6.5277   3.5854 -19.2911
#>   -2.0343   2.9953   5.9265  -1.3064  -5.3847  -7.5717  24.5642  12.3990
#>    2.7189   2.2196  -4.3728  -8.4464  -1.4428 -10.6433  -5.1407 -15.6691
#>    2.4052   3.5270  -4.3505  -4.1663  -6.4534   3.1089  11.7659  11.7800
#>    3.1321   9.0908  -1.7043   5.6446  -6.4308   8.2962   2.6201  17.4037
#>    4.7728   8.7987  10.7829   6.6520  -8.5370   8.6426   1.0484   4.7197
#>   -1.3202   1.4504   7.1640  -7.0080  -2.0330   7.0014  -5.6310   1.8968
#>    3.2068   2.2499   5.8587  -1.2549  -3.2135  -8.1138   2.5483  -0.1443
#>   -2.9971   0.7340   4.8800   5.0160   5.2999   0.5449   3.4508  -6.9330
#>   -3.6953  -7.9125  -6.5903  -1.6421  -3.7542   4.6884  -7.1345   8.2232
#>    1.5349   4.6420   2.3704  -6.9034   2.2848  -2.7961   2.9063  -0.8063
#>   -0.8225  -0.8009   7.5698  -0.2230  16.2769   5.9140  -4.3657  -5.3534
#>   -2.6549  -1.3497  -9.6197   2.3881 -14.9435  -0.1448  -4.3858   6.8263
#>   -3.3280  -3.5365  -7.6359  10.3882   2.1174  -7.8781 -23.9534   5.7035
#>   -1.8100   1.9984  -5.4845  -8.2628  -3.8468 -15.7736   3.9048   1.6131
#>   -3.7565  -3.5314 -12.8774  -2.1280  -9.6342  -6.2392  12.0370  17.4783
#>    6.3952   1.1655   1.7103  -2.7094   6.0874   6.5419  -4.6324  -1.1755
#>    0.3859  -2.1804   9.2676  -7.8195  14.5370  -1.5975   0.2085  -5.6324
#>   -5.7318  -3.2608  -1.9139  -6.8097  -1.7958  12.3060  -8.4478  17.3006
#>    1.1163  -0.7892  -0.3992   0.9056  -7.3320   4.5790   5.5438   3.5440
#>    4.0788  -0.7213   7.1785  -8.9343  28.7723   0.7641  -9.5899  -8.9140
#>   -0.1369  -2.6693  -5.2058   1.6916 -11.0870 -11.6009  -8.2888 -10.0680
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]