Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.8604  -8.1736  -7.5769   0.6553   0.6294   7.5804 -14.2729   7.1598
#>   -0.9931  -5.1487  -7.2111 -12.7569   1.1627  -2.0358   9.4720  -3.5948
#>   -4.6174   0.4121   6.8202 -10.5907  -7.1448 -13.3897  -6.7425  -3.8138
#>   -1.7121   4.2262   4.1266   5.4241   9.1962   7.5686  11.4756  -4.1692
#>   -5.0204   0.1404  -5.1104  12.6313   9.9349  -0.3226   3.0947  13.0819
#>   -1.2752   0.3608   3.5182  19.9581 -11.0117   1.8910   1.1732   9.5847
#>   -2.3293   5.0700   3.0396  10.5344 -11.5434 -10.5018  -2.4273   8.7773
#>    2.8376   1.2755  13.8922  16.6175  19.5126   5.1637   2.9879  -8.7207
#>   -0.3053  -1.5704  -9.0093  11.0019 -12.0409  17.3298  -0.7452   0.0042
#>    3.0096  -2.6004   8.4339   4.5911   8.2291 -17.1927  12.3489  -8.8437
#>   -0.5052   3.6561  -5.4422  -4.3393  -1.8903   1.2384  -3.8552   0.8590
#>   -7.9428   2.9890  -4.0533  -1.8180   0.7142 -11.0357  -1.2132 -10.9474
#>    4.0718  -9.0917   7.2333   0.9765   8.4781   3.9622  -4.2934   5.6191
#>   -1.9903  -4.2842  -2.7467  -6.1547   4.0394   5.1229   9.2323  12.5890
#>   -0.8866 -10.9298  -6.0796 -11.4896  15.6702  17.1530  -1.2733  -3.4750
#>   -5.3032   2.0882   7.4416   4.9286 -10.2894   1.6753  -1.5652  -5.9930
#>    2.0543   2.1938   1.1033   1.1819  -2.3211  -0.3455   3.3510   0.0230
#>    3.2369   0.2982 -11.2371   6.3580  -1.0647  -3.5850  -1.6225  -0.2644
#>    1.1842  -7.9533  -6.9662   9.0754  14.9404  -3.1781   1.7261   0.8176
#>    6.6895   3.3036   2.4033   5.9235   2.8829   4.5779 -15.8226   8.3654
#>   -4.8838  -2.8507  -5.7091   1.0318  17.2468   8.6458  -2.2148  -3.2713
#>    0.1808   1.0026   6.4621  12.1607  -5.2672  10.1829  10.5724   3.0600
#>   -1.5908  -8.6716  -6.6672  -6.6475  17.5940  10.9376   8.4384  -2.5359
#>    0.2812  -3.9224  -1.5815  13.6550 -15.1246   5.5861   1.7264  19.1196
#>   -3.0224  -2.6805  -0.5044  -9.2873  13.4311 -14.9028   5.1894  -5.9338
#>    7.3594   4.1636   0.4577   1.2523   3.1750  -9.9062   4.2880  -6.0081
#>    0.5771   9.7972   2.2456   9.0198  -5.4285  14.0656  -2.7503  -8.1487
#>   -0.1564   6.5573   6.5493  -1.0071  -1.4878  -2.1987  -0.4480 -15.9391
#>   -0.2171   6.3545  -0.1683  -9.0770  -0.8171  -8.6749   7.6966  -1.2764
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]