Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.5780   0.0083  -7.3074   2.0192   6.1297  14.7140  -7.0665  -4.6233
#>    0.7851   9.3110 -14.4249  -6.6271  -4.2305   2.6864  15.1751 -21.1597
#>    0.1834   0.4340  -5.7832  -6.3381  19.4366  -5.8897   5.9350   4.3535
#>   -1.7856  -4.5532   3.8093  -0.2957  -2.0407  -0.4882  11.1887   3.8086
#>   -3.3132   0.8454 -12.6437   1.6194   6.4123  -2.3333  -0.6341  13.7512
#>    1.7655   1.1860   2.5684  -5.5520   8.1877   4.8859  -4.1666   0.6706
#>   -0.0705   9.2356   4.5626 -15.7271   6.8026  10.7563   8.9125   0.7695
#>   -0.3683   2.1792   6.4228   6.2835  -0.9995  -3.6147   9.0533  13.6937
#>    5.3692  -0.6657   7.3491   4.4551  -2.9418  -2.0900  -8.8623   0.5788
#>   -1.2074  -7.2865  -8.1142  -1.6302 -16.6642   6.2893 -11.0911 -17.1138
#>   -0.6391  -8.2844  -6.6502   4.5864   3.5516   6.3427  -1.0996   2.3919
#>   -3.9482   4.1949  -3.4663   6.0508  -0.2631   0.4275   5.5206   3.5888
#>    4.1843  -1.0223  -2.0677   2.2361   0.2515   6.2475   6.4102   0.6956
#>    2.0577   2.2596  -4.7010  -4.2021   0.4735 -14.5299  17.6543   6.5176
#>    1.5887   0.9080   1.3305  -8.9817 -15.2211  -0.4326  -3.7310 -13.0399
#>   -1.4224 -13.8140  -2.8171  -3.0408 -12.9426   0.6397  -3.5394  -1.9503
#>   -1.3362   0.1777   9.5510   3.2373   7.7781  19.4556  10.5445  -1.4554
#>   -6.2807  -7.7299  -1.8185  -7.9040   7.4080 -10.9771  -0.2917  -0.7883
#>   -2.3315  -1.2977  -3.5753  -3.8668  -4.1762   8.3244  -5.8657   6.3374
#>    1.9201  -3.1680  -4.6449   9.9316  -1.2064  -4.1335 -12.8972   0.1831
#>   -1.2414  -5.3534   4.7027  -8.8660  -9.9597  -7.5438   0.6914  -8.6569
#>   -2.7963  -4.3414  -1.0523   2.4644  -3.9786   0.2169  -1.0684 -13.5520
#>   -0.4495  -3.2727   0.5581  -0.8451   9.3056   2.7897  -4.4817  -9.0905
#>    5.1088   9.2977   4.7210   2.2318   3.1533   8.3103  -5.5475 -10.8000
#>    0.3378   9.7512  10.1889  -0.1036   2.5595 -11.8865  10.2236   5.4483
#>    4.3847  -1.6399  -8.0041   0.6555   0.5864  -3.1361  -5.2024  -4.5901
#>    0.6270   4.0388  -8.4977   3.3904  -3.2071   0.8510   5.4300  -3.6765
#>    1.8584   0.8892   6.5852   0.9882  -2.1068  -2.1539   5.1478   6.6671
#>    2.6551  -9.0740   5.3143  -6.7008  -2.6333   4.1925   0.9781   0.5228
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]