Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.2673  -1.0607  -4.4616   0.9855  -1.5365   1.8637   0.0963 -11.9033
#>    6.3845  -2.2464   1.4599 -14.2434  10.3256   5.9164  15.7862  -5.8033
#>    1.8004  -4.6298  -8.0214   0.3055   2.5062  -1.3833   4.2933   5.0488
#>    2.5838   1.9263   1.4206  -2.0035   9.3307  -2.6036  -7.4029  -0.8578
#>    7.7760   1.7216  -2.5371   4.9584   8.7630 -16.8436  14.0242  17.0547
#>    3.2418  -5.7155   1.9244   7.7267  -8.1048 -12.7443  -9.4697  -5.5303
#>    1.8891  -6.4709  -4.2391  -7.8278 -18.3772   1.3564   1.5666 -11.7474
#>    1.5503  -5.0142   4.6909   9.4672 -18.7822   5.3728   0.5353  13.2044
#>    8.5291  -8.4874  14.2147  -1.0578  16.1755   8.7524  -5.2951  -3.4857
#>    4.4909   6.6492   0.0437   7.9525   5.5162   1.7737  20.5137  12.2615
#>   -5.4853  -7.2640   9.8884  -4.6499  10.5823  -4.9115   4.4013  -5.5311
#>   -6.9315   7.3292  -0.2161  -4.7320 -12.9895  14.5836  -5.3146   6.6231
#>   -8.5136   4.4158   4.3757   3.7198  -2.7305  -9.7873  -7.6271  -3.5509
#>   -0.8971  -6.9486  -0.9231 -10.4929  -9.9785   2.6045   9.2904  -0.7426
#>   -5.1210  -6.1715   1.9718  15.7923  -4.1163   0.7136   9.2525 -14.2971
#>    4.1338   1.3485  -0.8180  -2.2137  -3.2486  -8.0650  -1.4194  -3.8504
#>    2.8327   3.0351  -3.1611  -8.0578   8.0540   8.2347 -10.6665   9.7001
#>    2.2996  -2.2785  -6.0006   3.2093   6.1039   1.8196  -0.2187  -1.1278
#>   -2.5367  -6.9879  -4.8875  -6.8234 -10.2221   1.0453  -6.4421 -18.2084
#>    3.7873   7.6546   4.6719  -0.3959  -1.5939 -16.0406  10.0407 -11.6031
#>   -3.7450  -1.3959 -13.2623  14.3452  -6.3458   5.3652  16.8926   0.2719
#>    2.7037  -2.2112  -3.3327  -3.6657  23.9091  -8.2678   2.4831   9.8714
#>   -6.3294   7.3906  -1.7492   6.9173  10.7210   9.4397  -1.2422   5.1722
#>    4.6509  -2.8316   8.7929  -8.2515   6.3782 -10.1177  -0.5966 -18.1547
#>   -4.9458  -1.6476  -8.0334  -2.1828 -16.3696  -6.3485   9.7073   7.3524
#>   -0.5664   1.8745  -2.1849  -9.0517  -8.9111  -5.7218   5.5938 -11.8773
#>   -2.3134   4.4657  -7.9987   3.5911   3.6645  -0.9214  -5.4369  15.4729
#>   -5.8575   3.9988   3.6542  -1.4005  -0.0915   5.1352  -5.5597   8.6531
#>    5.0345   8.9677  -0.9575   3.5279   3.2344   1.0006  -0.1003  -2.0410
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]