Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.9923   8.0592  -7.8378   0.9494  -1.4260 -13.4453   2.5154  12.6778
#>    0.0805  -2.5933   1.8862 -12.0419  -1.9417  13.4690  12.2592   2.8178
#>    3.6967  -7.9024  -0.1036  -1.4899  10.0210  -8.1523   5.4224 -19.3404
#>   -3.1892  -0.7999  -8.4331   6.0229   5.5473 -18.9386  -6.4083   6.5250
#>   -1.3872  13.2064  -3.3491   5.0108  -6.5251   4.7386 -10.2142   2.6646
#>   -6.7287  11.0806 -10.5983   1.7435  -5.7964   2.2807  20.4334  16.6858
#>   -1.9118   2.1918   6.1742  -0.0375  -5.2174  14.9561  -9.4966  12.8938
#>    8.3218  -0.7840  -2.0567   6.8251  -2.9609   5.0961   2.1000   8.5053
#>   -4.1675   7.9971   2.7304   8.4140   0.4627   6.6835   4.9838   6.7454
#>   -5.3364  -2.2618  -3.7807  14.8305  -9.0794   6.3144   6.1959   0.4734
#>    3.3307   3.7284  -3.1116  -2.1438   0.6509   2.0049   8.8997   4.4551
#>    1.7805   3.3060   0.0769  15.8046  -0.5158   2.0221 -16.9749  -0.5655
#>   -9.0628   3.7760 -15.2317  12.0446 -14.6964   4.0533 -11.2268  -2.9099
#>    3.0538   4.7007  -3.1019   1.3059  -4.3936   3.6282 -10.8813 -17.2339
#>    3.9963 -12.7718  -2.5011   3.8076  -6.9387 -10.7588 -24.0635  12.5136
#>   -5.6171   3.8246  -3.3090   0.9685  -0.7664 -15.7695  -7.1327  -1.5463
#>   -0.3215   0.2566   0.5014   1.2598   9.5064  -9.5120  25.3924  -1.7116
#>    5.5681  -4.8272   5.4008   3.9059   8.1687  14.2034   8.4243   4.9449
#>    5.5882 -10.9426   6.0786  -9.1714  -1.0778   6.0958  -6.4361  -7.4874
#>    1.9519   3.5286 -16.2327   6.1439   1.6346 -12.5406  15.3072   5.4721
#>   -6.1573  -1.5919  -0.7614  11.5464  -8.7256  -1.4751  10.5468   3.7963
#>   -0.8845   5.5291  -0.0042   1.1893  -1.1051   8.0944  -1.1057  -4.4694
#>    2.8535   2.7004   3.6939 -10.4503  10.9665  -9.8903  -5.9991  -1.1271
#>    0.5153   3.5268  18.7494  -1.9124  -3.3307   8.9416  -3.6819  -2.7654
#>    1.2453  -3.7570   4.2943  -6.8246   2.1816 -15.1042  -4.9414   2.0134
#>   -1.6349   4.3060  -1.5626  -3.0125   0.5952   9.4915  -1.2651  18.7714
#>    0.9445   1.9758   2.1252   0.1831  -6.7248  -5.4268  -7.6814  -3.4332
#>   -0.5034   3.3590   5.2299  -2.1954  -5.1127  11.6811  -1.8035   6.1038
#>    1.5539  -7.2978   1.4877 -13.7847  16.5904   7.0102   3.0947 -17.9150
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]