Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.5322   0.3675  14.9496   4.8312   3.1642   5.4207   9.9953  -8.5996
#>   -4.0218   0.6073   3.0469  15.9226   6.6730   5.1216 -14.9858   3.1880
#>   -7.8350   0.4432  -0.8683   1.3062 -12.6362  12.3847   3.7136  15.8189
#>    7.3732   2.0548  -0.6872  -8.6476  -0.8624   0.6673   4.8813  -7.8423
#>   -3.8436  -1.3541  -0.7543   5.8280   3.2773   6.8661   6.3879   2.2113
#>   -3.7626  -6.2821   3.5305  15.8960  -6.7862   1.2675  12.4381   7.1319
#>    0.9586  -4.3763  -5.3494   4.2185   2.7042  -2.8318   2.4894   1.9049
#>   -8.0521  -5.7474 -14.0530  -3.5326  -4.0057   0.1441   2.1657   4.6690
#>    5.0618  -0.2756  12.9648  -7.9776   0.7868   2.8056  -0.6336  -1.6795
#>    5.4669  10.5946  -6.6020  -4.5978   0.3307  -1.2305  -7.9330   5.5765
#>   -9.2677  -2.8956  -5.2612   5.1022   7.8135   5.0373 -14.2047  -0.2965
#>    1.7197  14.1518   2.6641   5.4847  -0.6492   8.4323  -4.9216  19.6715
#>   -2.6961   0.4285  -3.9537 -12.3836  -5.0374   1.9621 -15.1084 -35.8487
#>   -2.0959  -9.2844   1.4160   5.6460   9.4832  -2.9867   3.1385  -3.8410
#>   -0.7480  -5.5188   0.7743  -6.5721   1.6626  -8.2913  11.1819   6.7042
#>   -1.3516   1.1908   0.4920 -10.8242  -0.9862  -5.4730  -4.2911  10.6957
#>  -12.0205   0.8879  -2.5191   1.4909 -16.9131  -0.5142   4.5119   0.9514
#>    3.6733   6.8786  -6.0918  -7.1850   8.1831  -2.3837  -3.1015 -10.3027
#>    3.5052  -0.2488   5.4610  -5.1263  -2.4094   0.4315 -10.0822   0.7403
#>   -1.8578  -3.9228  -1.4982  18.1571   1.1239  -8.4727  -9.4639  -1.9554
#>   -1.7443  -7.0505   4.9250 -12.2512   5.5580   2.2336  -1.9354 -15.6150
#>   -8.3528  -2.1246  -0.8153   2.5804   7.4730  -1.3800 -16.9917  -7.2608
#>   -7.6821  -2.3231  -0.3699   4.4170  -4.5075  17.1298   6.6493  -2.9460
#>   -1.4862   9.3287   4.2491   3.0937  14.5515   4.4475  -4.8122  -8.6457
#>   -8.8873  -5.1940 -15.5853   8.1903  -6.4801  -0.4712  -2.7259   1.1635
#>   -4.5782   5.7440  13.7405  -3.7777  -2.0211  -0.8241   4.4528  -9.0544
#>    4.9324   2.2437   6.7466  11.7283  -8.4042   0.0072   4.5767   5.5788
#>    8.3271  -0.9222   7.8477   5.2628  -2.0168   7.2704  -3.1129  -3.7836
#>   -8.4048   2.5446  14.2834  -0.3862 -13.6873 -14.4278 -11.6925  10.2362
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]