Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.8505  -3.7832  16.3873  -4.0731   4.3106  -2.3081   0.8601  -4.2873
#>    7.0348  -0.4539  -1.9219   8.9910  -7.0407  -8.9272  15.1111   8.9325
#>   -1.2001   1.9726   2.6372  -2.9832  10.0094   9.9711  -3.3104   2.5701
#>   -0.4011   4.6514  13.3235   6.6504  -9.5786  -6.3485   7.5897 -12.5879
#>    2.1117   6.4094   7.3998   3.3195 -17.9599  22.4149 -15.3775 -15.7385
#>   -2.8490  -2.9081   9.5393   6.4464   4.0963   0.1399  10.2931   3.7793
#>    5.6934   0.7375  -2.5902  10.8487   7.2124  -3.8535  17.7747  -5.2110
#>   -1.3798  -8.3504  -5.2716   7.1629   0.1536   5.9955   3.3708   5.7489
#>   -0.0143  -0.4330  -5.6982   4.0959   4.4014   5.6736  -9.1871  -9.1652
#>    0.9508   5.4821  16.3350  -6.4846  -4.0534   4.7556   3.1864  -7.6487
#>    3.8597   4.9622  -2.8964   4.8386  -5.5665  12.5836  -7.3919  -4.3919
#>   -2.0788  -7.6024  -2.2394   4.7675  19.2459 -12.5131  -9.6350 -11.6968
#>   -3.3363  -6.1951  21.7714 -15.6366   3.5687  18.7434  -9.6826 -14.9482
#>  -10.0617 -17.5923   9.6715  -0.9530   5.3641 -10.7534  -5.6614   2.2345
#>    0.1539  -0.0303  -5.3105 -19.2250  -1.0916   2.9125  -1.2321   8.6005
#>   -1.6588   1.4496   8.6691 -13.5419   6.8522   0.6148   0.3831  -6.7409
#>   10.8267   5.5922 -13.1280   1.6407 -23.2989   9.9964  -3.3360   0.8102
#>   -4.2903  -8.7740 -15.6098  -3.8153  20.4282  -2.1366 -11.8353  12.3347
#>    0.5120   0.2740  -0.5521  -7.5554  -3.1433  -2.3304   3.7636   9.8791
#>   -6.8557   3.7126  -0.6932  -3.9655   9.0300 -10.7287  -9.5113   7.5320
#>   -0.5202   9.4168  -5.7316   9.4124 -17.5814  -3.2889  10.7718 -11.4720
#>   -1.7890  -4.4130  -4.4426   5.6488  -6.2644  -0.5131  -4.4196  -5.0676
#>   -1.3555   4.9118  -1.8601   7.7721  -9.6897 -15.0980   2.2436   5.0464
#>    2.2692  -4.8317 -14.0634   4.3425   7.1924 -10.1236 -24.4072 -13.3116
#>   -5.7077   1.4924  -9.0756  -9.9988 -11.8735  -0.7244  -7.7062   0.7495
#>    2.4291   0.6225   3.9365   2.8142   8.6791   6.6603  -0.7976  -2.4579
#>   -0.2973  -1.2547   6.7680   3.0009  -4.7618  -5.4344  -7.5976  11.8722
#>    1.2299  -3.0382  -3.7191 -11.7297   2.4421   7.3118   6.2071 -10.0324
#>   -0.9724   6.4954   8.7069  -5.1840   5.4352  15.5068 -16.7987  14.8335
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]