Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  4.2567e+00 -2.3242e+00  4.1790e+00  6.4722e-01 -3.2965e+00  8.2855e+00
#>   3.9487e+00  8.2872e-01  7.0519e+00 -2.5529e+00  1.6571e+01 -2.0003e+01
#>  -6.3354e+00 -6.1886e+00  8.0075e+00 -2.4746e+00 -2.2853e+01  8.7423e+00
#>  -4.6940e+00  5.4468e-01 -5.1507e-01 -4.8792e+00  1.3490e+01  5.5015e+00
#>  -1.8182e-01  1.3629e+01  9.0157e-01  8.0834e+00 -6.8777e-01 -6.6273e+00
#>   7.0145e+00  4.8539e-01  1.2269e+01  2.9918e+00  6.7189e+00 -1.6475e+01
#>  -8.4236e-01 -4.5305e+00  6.6884e+00 -5.2616e+00 -1.0762e+01 -3.3126e+00
#>   2.4969e+00 -3.1036e+00 -1.0555e+01 -3.4916e+00  2.6037e+00 -2.5810e+00
#>   4.7608e+00 -8.7726e+00  4.7969e+00 -1.3526e+01 -1.4690e+00 -6.3307e+00
#>   6.4985e+00  8.5778e+00  6.5644e+00  1.2625e+01  1.8790e+00  4.3301e+00
#>  -5.9909e+00 -1.5737e+00  6.2147e+00 -9.6791e+00  1.1895e+01 -4.6301e+00
#>  -3.0552e+00  5.3279e-01 -3.7013e-01 -4.7407e-01  1.1933e+00  1.7746e+01
#>   2.0132e+00  6.8690e+00 -1.3725e+01  1.4463e+01 -5.7460e+00  3.3707e+00
#>   4.4982e+00  5.0654e+00 -2.0461e+00  7.2227e+00  1.3940e+01  2.3545e+01
#>  -9.6261e+00  3.0613e+00 -1.1356e+01 -1.5448e-01 -9.2094e+00  1.0705e+01
#>  -2.4670e+00  2.8456e+00 -9.3712e+00  1.0549e+01 -1.1815e+01  3.6074e+00
#>   6.0229e+00 -1.0911e+00  4.5797e+00 -3.6777e+00 -2.2890e-02 -4.1238e+00
#>  -2.3312e+00 -2.9663e+00 -1.4921e+00 -2.0303e+01 -1.6536e+01 -5.5778e+00
#>  -3.5282e+00 -3.1360e+00  8.2843e+00 -5.5185e+00 -8.8877e+00 -6.0619e+00
#>   2.8285e-01  3.0613e+00  8.8730e+00  1.0556e+01 -1.2085e+01  3.1028e+01
#>  -6.2525e+00 -1.8608e-01 -2.5838e+00 -9.4960e+00 -7.4262e+00  7.0940e+00
#>  -2.0510e+00 -2.9298e+00 -4.7448e+00  8.3569e+00  8.5500e-01  1.5249e+01
#>  -1.4889e+00 -2.7842e+00 -2.2785e+00 -3.1258e+00  2.5865e+00 -2.9822e+00
#>  -2.0739e+00  6.0881e+00  1.4922e+01 -6.0696e+00  9.5848e+00  1.0489e+01
#>   7.7878e+00  9.4007e+00  7.1663e+00  5.7329e+00  3.1741e+00  3.4690e+00
#>  -1.7542e+00  1.5489e+00 -6.1094e+00 -4.9879e+00  7.4211e+00  1.6199e+00
#>  -2.8481e+00  2.1085e+00 -1.5118e+01 -4.7813e+00  1.4163e+00 -8.2588e+00
#>   1.0656e+00  7.5397e+00  5.0277e-01  7.3645e+00 -5.8692e-01  4.0798e+00
#>  -1.1601e-01  3.0216e+00 -1.4293e+00  1.8252e+01 -1.1025e+01 -1.4728e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]