Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.9678   8.4607   2.9684  -2.0617  -1.8444   0.9249  -3.4788  -3.2257
#>   -8.0307  -8.4614  -0.6178 -16.2459  12.4627  11.9146   4.8093  -9.8110
#>    0.7707 -14.3567  -0.9171  25.2675   4.0336  -6.0270   7.7935  -0.7070
#>   -3.0018  -7.2476  -1.0303   2.2621  -5.4289   4.1791  24.2895  -1.5052
#>    2.4850  -1.3366   8.3158  13.7658 -10.0913   4.8103 -12.3035 -14.4365
#>   -0.7901  -9.5412   5.9508  13.1689   1.5658 -21.1939  -5.2130 -20.0421
#>    0.5151 -14.3041 -13.8165  -4.8987  10.1523  -3.1150 -10.6837   5.7317
#>   -4.1229  -7.6820  -2.8248  -0.0479   4.5292 -14.5050  -2.1598  14.1848
#>   -4.1398 -12.0796   1.0841  -7.1734  12.7290  12.1557   3.3097   3.7426
#>   -3.7064  -0.4855   1.0616   0.8998 -17.3223 -22.3545   8.1111  -2.9086
#>   -2.4070  -7.5039  -5.0826  -8.3426  -4.5446   8.9173  -0.6346   4.0312
#>   -5.2045   1.2306   1.3853 -10.1149  -2.7141  14.8220  -7.3880   6.1112
#>    7.2435  -2.4487  -0.3811   6.1587  -5.1669   5.9424  -2.7540  -4.2349
#>    5.4743  19.0113   7.7234   7.9583  -9.5776  -0.2809  10.7821 -16.8390
#>    3.1252   1.8024  11.9958   7.6281  12.0178  -3.0605   5.9765   8.6728
#>   -7.2135  -2.3719  -5.3796   8.0302  17.6362  10.5587   5.7996   6.5687
#>    4.5150   8.1998   5.6032 -14.3897   7.8245  -7.6610 -15.9103  10.0822
#>    3.5053   3.6682   9.9336  11.1983   1.0618   7.2014   0.1165 -18.8494
#>   -2.1325  -4.0537   2.4340   7.7719  20.3017  11.8472   5.5848   0.0504
#>   -4.4347   2.1491  -2.7229   3.8217  -4.5675 -12.9434   9.6906   5.3705
#>   -2.0464  -6.4367   0.2942  -5.2225 -10.9162  -0.7262  -0.8927   9.9095
#>    3.5897  -4.7293  -9.9500  -1.3818 -14.9491  11.6942   4.2731   8.3041
#>    7.9069  -4.7511   0.8460  -2.8899  19.3368   0.4504  -7.3574 -10.8131
#>   -1.5813   1.3098 -13.2993  -1.4653  -3.1573   9.0724  13.7154  -3.2085
#>    9.4560 -10.9786  -5.2639 -19.8230   2.8178 -11.5536   1.4828   1.9408
#>    1.4407   9.2986   2.1405   1.5781  12.9713  11.5707   2.5429 -10.2850
#>    3.3467   2.0206  -2.0506   3.0244  14.7811  -9.3717  -5.9723   3.1504
#>    0.3669  -6.8521   3.0079   0.0629   1.7361  13.2719   0.1257  -2.3434
#>    4.2173  13.4703  -6.2424  -2.3254  -0.1144  -3.9085  -2.6769   2.8584
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]