Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.5739   1.0379   4.0855  -0.6329 -11.0282  -1.0811 -12.2802   2.8251
#>   -1.5817   3.7719   4.1146   4.4118  -0.6182   9.8201  -2.3889   5.2739
#>    3.1749  -3.7740   7.7322   1.5450   1.0324  -2.9328  -9.0378   8.3817
#>   -2.1253  -1.7312   9.0076   5.1860  11.9184  20.5515   1.3127   3.8768
#>    3.6266   6.8126  10.0541  -6.8443   2.6801  11.3834   5.0416   5.8411
#>    1.7200  -0.4041  -0.8699  -9.0800 -10.4856   8.9249   2.3711  -9.7765
#>   -0.9331 -10.8440   8.9700   5.6427  20.8118  18.0635   6.2120  13.6781
#>    2.8638   0.4787  -5.8761   1.2388  16.0788  -5.9775  -2.9099   3.6601
#>   -6.8558   6.3397  -5.9424  -7.9293   2.0471   3.3951  -0.9000 -10.4611
#>   -4.7319 -10.5738   5.1176  -5.0613   4.1610   2.3778  -6.6794  -0.8086
#>    4.1323  -9.1329  -2.6424   3.1701   2.8869  -1.9248   6.1310   5.4588
#>    6.1690  -4.6313   5.1366   3.9958  -8.5884   4.0944  -0.2572  10.1216
#>   -2.3090   8.0133  -3.6032   4.8633  11.2083  -4.0590   2.8482  17.0815
#>    0.2338  -4.1732  -2.8948  -8.7087 -11.2162  11.7972 -12.5906  -5.3018
#>   -1.8101  15.0014  -7.3779  10.1147   2.6569   3.6257   3.8964  -3.0928
#>    9.7760   6.1010  -5.6808  -3.5362  -3.6124  -3.8934  -7.2418   3.8170
#>    3.4429   1.5667  -1.2854  10.9003  15.7763   0.8993   4.5116  -2.6750
#>   -1.7394  -1.5147  -7.0184   4.4300  -0.3472  -1.2805   0.3371  -4.4634
#>   -1.5574  24.5594  -2.2535   8.0158  19.4341  -8.7420  -5.9973   6.1831
#>    0.8450   0.7366  -0.8324  -1.1222   3.3044 -10.0323   5.5291  -0.4113
#>    1.5348   3.4223  -4.6355  -5.7673   4.3616   8.6587  -3.0999 -11.6885
#>    0.5504   3.4696   1.4740  18.7138  18.0663 -13.1735 -17.2088  -0.2546
#>   -2.1918   6.1384   1.2416   1.4815  -3.1083  -3.9624   3.7754  -1.8864
#>    5.5375 -10.2393  23.4397  12.7052  -9.1100   8.3985  11.3133  18.1012
#>   -3.5946  -2.7632   1.6770   7.6248 -18.9728 -14.5905  11.5571  -9.0759
#>   -2.7910  -0.4425  -2.8880   8.0055   2.0065   0.4372   1.0510  -1.1247
#>   -1.6470  -3.2955  -2.0887   4.2174   1.7460  -9.6168   3.3176   8.8833
#>    0.8624  -4.3468   2.8088  13.5437   1.7156   8.0787   4.6640 -14.1465
#>   -1.0035  -3.1880 -10.8066  -4.1708   3.1036   4.3286   2.3839 -16.3603
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]