Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.3182  -1.7871  -0.4836  -1.3775  -9.2185  -1.9718  -1.3740   4.7814
#>    0.7512  -2.0525   0.0306   2.8051   8.2856   0.2937  -3.5275  12.6141
#>   -2.7042   4.1064  -0.8162 -11.4798   6.5779  -8.4973  -6.9293   6.9960
#>    1.0749  -3.2424   6.7650 -13.3418  15.0123   2.0331  -6.3854 -17.7225
#>    2.7690  -7.6922   5.9494  -5.3147  -0.5979   9.0924 -16.7270  -5.8146
#>   -5.4685  -0.4139  10.1136 -11.4878   1.9210   2.5730  -6.4277  -7.8489
#>   -1.0934   1.8936   4.6883   1.9000   6.5194  -3.9876   3.7262 -11.2110
#>   -2.2029  -4.0507   5.7324   4.3535   8.9093   4.4864 -21.6393   0.6864
#>   -3.3127  -1.1386 -15.7759   7.2119   1.2549 -10.3084   8.4850   6.2223
#>   -4.6317   7.0488  -8.8821  -0.9842  17.0392  -7.3905   2.7969 -22.1863
#>   -0.3465  -6.5515   6.6390  -4.2257  -1.6487   5.5710  -1.4736 -15.7769
#>   -3.4137   1.5834  -1.4295   0.7186   6.4910  -5.2893  -3.4365  -4.6107
#>    6.4125  -2.7780   0.0191   2.2386  -0.1145   8.5384   2.3448   0.0853
#>   -2.1919   2.5447   9.4478  -1.9937  -6.4892   6.6642   2.5779  -2.6407
#>   -2.0157   4.4176  10.0786  -0.0833  14.8119  -7.2459   0.5267  14.0071
#>   -1.4718  -5.6641   2.7921   7.4159   5.1286 -13.8192  21.6966  -8.7590
#>   -1.8908   8.3877  -3.0814   5.7759  -3.4351   2.1645  16.5424   1.0739
#>   -1.0697   2.9148   3.2008   1.1393  -5.5737  -6.5946   7.6071 -17.7486
#>   -1.2939  -7.6436   9.4855  -1.5881 -12.0493   9.0064  -0.1258   6.5516
#>    3.0315  -5.1177   3.2270   9.2329  -4.8231   8.5031  17.5864   5.6215
#>    6.4100  -5.8050  -6.2843   1.9183 -18.5548   5.3152   9.4898 -23.0072
#>    2.3838   1.1795   4.8074   1.3575   5.3984  -0.8054  -7.8845  -0.1047
#>    1.1957  -0.0601  -0.9868   1.2375  -0.4664   0.0404  18.1882  15.2964
#>    0.1861   0.9407  -3.8204  -5.5289  -7.4195   2.2884  26.5084  -3.2186
#>   -3.0097  -5.6450  -4.3902  15.1278  -3.4745  -1.9421   9.7448  -7.2852
#>    3.0939   0.2652  -0.0045  -0.9602   1.2139  -9.3888   0.7835  -3.4919
#>   -2.6903   2.6397   2.8204  13.1967  -3.2962  -0.0664  -0.7266   5.3542
#>    1.7021  -7.5616  -1.9225  -9.4640  -3.2200  10.6057  -7.6587  -7.8765
#>    1.3010   2.1384  -6.9501   4.0957 -20.4881   5.4334  -4.7090   5.2120
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]