Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.5124  -4.4562   1.7391  -1.7565  -9.0602   3.9010   8.6758  10.1091
#>    0.3254  -2.6307  -3.9885  -2.0496   2.9142   0.7026  -2.3610   3.6041
#>   -2.0030   1.1535  -6.0192 -11.4633   9.7466   7.9003  -3.8225   1.8081
#>    2.4727   0.5840   6.4032  10.9767   1.2029 -10.6465 -17.0789  -1.6882
#>    0.4310  -8.1351  -3.7333   2.3332  -5.8034  -1.1277   3.8360   2.9205
#>    8.1791  -3.1638   3.1978   4.4116  19.3100 -10.2394  -4.7267   6.7511
#>    7.0285  -4.1921 -11.0425  -2.0307   0.1307  -6.0147   4.3977  12.5442
#>    3.8951  -7.0152   6.1605   5.4262  -8.7458  -7.4148 -11.2938   6.9995
#>   -1.3883  -0.8161   0.9642   0.1712  -3.0130   0.0127  -3.8318   5.6475
#>    1.2993  -4.2216   5.0970 -12.1658   1.9788  23.8276   7.2231 -13.8344
#>    1.5215  -5.0423   0.1651   1.2448  -0.1203   8.1918  -0.7536  -6.7876
#>    3.2219 -10.7449   9.5893  18.3362   5.4531   1.8707  -8.4788   3.4241
#>    1.9811  -1.9771  11.1577   5.7603   1.6240   7.0275 -10.0636  -4.7840
#>   -0.7713   0.2847  -2.8469  -2.0393  -2.8222  10.0019   5.5954 -10.4623
#>   -5.3847 -12.2128 -12.6412  -8.9743   6.2300  10.8576 -10.3612   7.9906
#>    1.7157   6.8858   6.8509 -15.8846   9.2563   0.6232  -7.0836   0.2773
#>   -3.3834   9.5011  -2.0479  -4.8647   7.1539   2.7704   6.7326 -12.3262
#>   -5.7393  -0.6810  -3.1137   5.8760 -13.4917   1.6277  -5.8570 -22.9108
#>    5.6099   5.5790 -16.2623   2.1274   2.3460 -19.6314  12.7117  11.2198
#>    0.6961  -7.1109   0.2069   7.0558  15.3753  -2.9614 -22.7176  -0.8161
#>    2.7308   7.0552 -23.2863  -7.1579   0.9476   4.8896  -5.5864   3.6059
#>   -0.3688   6.8381   0.9325  -5.7053   0.9173  -9.1654  -9.4254   2.6790
#>    3.2255   6.5707  -2.9375  -9.5545  -6.0787  10.8867   3.3259  -5.2199
#>   -0.7561   2.1646  -3.6773  -3.7084  -2.9308  -4.1296  -0.0967   5.9654
#>    5.1601  -3.8329  -6.5495  -3.0396  -4.0345   0.2892  -6.0747  -4.2730
#>   -0.5822  -3.0210 -12.9838  -7.7168  -0.8251   6.0454  11.3786  -0.9797
#>   -3.9612   3.6554  -0.0673  -0.9563  -1.4641  11.7578  -4.5009  -8.8369
#>   -0.8022  10.0413  -2.2337   1.0689   3.1255  -3.8031  -3.6627   5.8794
#>   -0.1248  -2.1945  -0.2735   1.3114   6.9548  -6.4229  -3.2500   3.3538
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]