Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.9268   6.3275  10.8043   3.4831 -13.2939   7.6056  -1.2904 -12.5999
#>   -2.5078   0.7866  -4.1390  -3.4678  -2.9686   1.6258  12.7119 -15.6225
#>   -4.1389  11.3805  -0.6558  13.2667  -3.1733  15.1734 -13.8031  -4.6178
#>   -4.2883  -7.6041  11.4927  -5.5975 -12.3339   1.3235  12.5663 -17.5267
#>    3.6781   4.3686  10.6359   0.0386 -16.1445   3.6030 -16.9579  -1.2740
#>    4.1320   4.7316   7.5649  10.0369  -2.6030   0.0227   5.8050  16.3608
#>   -7.5975   2.0286 -14.6617  12.5185  -2.9501  10.9543 -11.4193   8.7821
#>    2.2242   2.7980   0.8116   4.1227  18.4479  16.1559  -2.7552  -5.2490
#>   -1.1783   5.1933  15.9243  -4.3552  14.0843   0.1687  -1.5340  12.4984
#>   -2.2089   5.1756   0.9259  -7.9842 -17.5338  13.3068  -4.4691  -7.5620
#>   -2.1182   3.2167  -3.6772  18.1626  -7.4829  -7.2505   0.6142   9.0007
#>   -5.5465   9.3553  11.7089   7.7582   3.8004  -5.2934   1.1095   4.7841
#>   -5.6277  -1.8238  -1.3565 -14.9055   2.2647   1.0816  -5.9662 -13.5600
#>   -1.4201  -3.0130  -2.9882 -16.4788  -6.7583 -11.1751 -15.9138   0.4057
#>    3.3046   1.5164   2.8509  -6.1587  -0.7391  -4.1533   7.2005   7.7902
#>   -1.9545  -4.0603   1.3193  -1.6321 -11.0809   5.4174 -12.4382 -12.7233
#>   -5.7394  -3.6763  -9.4566   2.9007   3.4361  -6.0765  22.8438  -3.5012
#>    7.7363  -2.2641  -4.3031   1.7991   1.3618  13.9752 -13.3231  -7.6600
#>    1.0956  -2.4254   5.2708 -14.4087 -18.7970 -13.1132  11.2015  -0.1379
#>   -1.8238   3.7353  -3.9332  11.1711  -6.7840  -4.5558   0.1212 -21.6073
#>    0.8439  -0.1116   1.5749   9.6409  11.5704   0.3999 -13.1134  24.6740
#>   -4.8170   0.5963   6.9925  -3.6541  13.0060  22.7920  -5.6401  -6.0863
#>    7.5544   6.7925 -12.1141   7.2471 -11.3649 -16.9358  -4.8935  -0.9186
#>   -1.5036   7.0058  12.3097   7.6362  -4.5320 -12.4987  15.9581  -5.4458
#>    7.7666  -1.5607   0.5898   5.1706  10.3612   7.2791  12.5363   0.7892
#>    0.0638   0.8082  -1.5146  11.1627  10.2516  16.7996   7.4419  -5.4058
#>   -0.7589  -0.6283   2.0059   4.5116  23.2823  -6.9666   5.4948  -5.2111
#>   -3.4568  -1.1072  -8.4831 -14.6211  -3.6808 -16.5385  -2.3173  -7.1791
#>    3.5052  -9.9866  -1.4124 -17.0272  -1.2329  -3.9354 -13.9656  -6.7960
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]