Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.5826  -8.5802   5.8864   0.1456   9.4506  -5.6767 -11.9226  -6.5477
#>    2.8149  -8.0258   2.4014  -0.6309  -3.5077  -7.9022  -1.6272  -1.8620
#>   -0.7002   2.1875   9.5624  -1.4587  14.2818   4.5730  -2.9332  -0.5629
#>    7.1310  -4.1671   8.6045  -2.9012  -4.5417  -5.1517  -1.2940   2.5803
#>   -0.1204   3.9759 -12.0908   5.8710  -3.0213  -3.5512 -11.1334   4.5401
#>   -2.7121   5.0667  -4.0863   0.2322   3.6473  10.7497  -1.9319  -7.9134
#>    1.0018   3.6633  -9.2299   9.1474   1.5304  -1.3295  -2.9789   1.2905
#>    0.2703  -9.5481  -1.6077  12.2357  -8.0255  -7.4578  -9.2565  -2.9073
#>   -2.7884  -3.2971   0.8397  -9.4439   6.3831  -8.6267  -2.4640  -4.1630
#>   -0.6816  -1.5965  14.7736  -3.3923  -1.6663  -1.5238   1.3198   5.1807
#>    3.7317   0.0899   1.7406   4.9787 -10.9071  -2.0837   0.1196 -11.5987
#>    4.1196  -0.2939  -4.8154   7.3814   0.0779  -6.5585   2.9769  -5.0025
#>    1.5369   8.2848 -11.2943   5.1032  -4.3229 -23.7106   8.9424  -7.2541
#>   -1.8383   5.8859  -6.6324   1.1159   6.4661   8.6036  -3.9639  -5.2235
#>   -6.0802   5.5682  -7.2483  10.9598  -4.8842  -4.7226  -5.3791   3.1313
#>   -1.0356  -1.9933  12.7585  -5.0809   2.9286  13.8742  10.5152 -12.5745
#>   -2.4134  -2.0552  -5.7414   0.8849 -10.6057   6.9884 -13.9924  -4.4721
#>   -4.7316  -0.5183   7.2213   3.0469  -7.0253   0.5661  -5.8826  -0.0830
#>   -3.1188  -4.4511  -1.0395   1.9318 -13.8229   2.1444  -8.0637  -7.6851
#>    0.2586  -5.6950  -4.9674  -8.8460   7.0801  -1.7118   0.8405  -2.6270
#>   -3.1629   9.6266 -10.3412   0.3329   7.5406  -1.8087   9.2912  -1.3920
#>   -5.5667   1.7146   6.2127 -11.6483  -1.1965   1.0573  -2.1709   4.1375
#>   -3.0689   1.6215  -1.8556   3.6326  13.5326  -2.2452 -11.8573   8.6016
#>   -0.4096  -2.3762   5.1201  -3.4837   1.0949  -3.3624  -5.5670  -2.3547
#>   -0.3453   8.0688  -9.4480  -2.1806   3.4933   0.1824   0.5429  -9.3317
#>   -5.9820   3.0203  -3.5316  12.2250  14.1229  -9.3228  -2.6051  13.3688
#>    2.5755   4.7827 -11.3001  -1.8084 -19.9100   5.5940  -3.9678  -7.5003
#>    3.2419  -2.0512  -8.2492   1.5237  12.0872   2.7935   3.9468  -8.3970
#>    0.8398   3.6504   1.0633   7.6643 -11.5981 -13.3934  10.5948   1.7297
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]