Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -1.3967e+00 -3.0816e+00 -3.7808e-01  9.8138e+00  1.5288e+01 -2.9202e+01
#>  -1.7654e+00 -5.1008e+00 -1.5515e+01  7.4194e+00 -2.6835e+00 -1.5969e+01
#>  -6.7090e-01 -1.2085e+00  2.7393e+00 -1.2427e+01  3.0191e-01 -1.1842e+01
#>   5.0587e+00 -6.0343e+00 -1.6711e+01  1.5552e+01 -1.3978e+01  6.8111e+00
#>  -8.0118e+00  1.0166e+00  8.8066e+00 -1.5109e+00  5.6156e+00 -1.3543e+01
#>   5.5605e+00  6.4688e+00 -9.2543e-01 -1.7893e+00  1.5169e+01  3.4224e+00
#>   1.0183e+01  8.5913e+00  2.4748e+00 -3.6688e+00 -3.3773e+00 -4.2997e+00
#>   1.7283e+00 -1.1997e+01 -9.2418e-01  6.3078e+00 -9.7801e+00  4.3932e+00
#>   5.0258e+00  4.3319e+00 -5.6916e-01 -1.9695e+00  4.3226e+00 -5.3142e+00
#>  -3.1089e+00  1.7566e+01 -1.1689e+01  5.7661e+00  5.9971e-01  2.3512e+00
#>   1.3179e+01 -2.7001e+00 -8.9951e+00  3.5681e+00  1.7234e+01 -1.5717e+01
#>   7.5190e+00 -4.9480e+00 -7.0309e+00 -6.2787e+00  6.5216e+00  1.3157e+01
#>   1.1142e+00  5.3935e+00  5.8881e+00 -1.0925e+01 -2.7318e+00 -4.5002e+00
#>   4.6911e+00  4.1643e+00 -6.4106e+00 -1.4631e+00 -1.5494e-01 -7.7579e+00
#>  -7.4792e-01  2.8169e+00  4.3780e+00  3.3686e+00 -6.2939e-01 -1.5637e+01
#>   1.4148e+00  9.8598e+00 -3.8213e-01  1.7742e+01 -1.6286e+00  6.3695e+00
#>   3.8409e+00 -3.6555e+00 -5.2972e+00 -6.6340e-02  1.3698e+01 -3.1905e+00
#>  -1.2319e+00  1.0027e+01  4.0187e-01 -9.4089e+00  7.9079e+00  3.0000e+01
#>  -2.3599e-01 -8.9998e+00  1.1974e+00 -6.3551e+00  1.1557e+01 -8.8522e+00
#>   2.1012e+00  1.0767e+01 -1.6818e+01  4.5603e+00  1.1210e+01 -5.9537e+00
#>   1.0724e+00  8.5789e+00 -5.7768e+00  9.4973e+00 -7.3010e+00  8.8500e+00
#>  -2.3589e+00 -5.2782e-01 -1.7333e+00  6.9181e+00 -1.0418e+01 -9.6310e-01
#>   2.1525e+00  1.7243e+00  1.7292e+01  1.0906e+00 -6.8870e+00 -3.3840e+00
#>   2.8187e+00  4.2469e+00 -2.1133e+00  8.3318e+00 -1.1535e+01 -7.1053e+00
#>   1.8499e+00  2.6256e+00 -6.5743e-01  3.3705e+00 -2.2846e+00 -2.0647e+00
#>   3.2761e-01  2.8161e+00 -3.9874e+00  2.6998e-01 -7.3103e+00  9.8540e+00
#>   4.4419e+00  3.0399e-01  4.9428e+00  7.0335e+00 -8.9518e+00  9.5338e+00
#>  -3.1428e+00  2.1450e+00  7.4036e+00  4.9109e+00  2.6541e+00 -1.6559e+01
#>   3.0636e+00 -6.8345e+00  1.5610e+00 -7.6870e+00  2.1337e+01  9.1268e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]