Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  10.8844   4.3183  -5.2986   3.2177  -1.1162   7.4137 -14.3993   4.8960
#>    6.2180 -14.4669   0.4141  -5.1050  -5.1913 -15.2110  -8.5214  -1.9352
#>   -7.8428  -3.1265   2.4695   0.7568  -4.9648  -9.8506   3.1634  -1.8980
#>    3.1316  -2.5861  -6.3654   9.4491 -24.1836  10.4751  -1.7318  16.8571
#>   -3.5502   8.0203  15.7600   2.7368   3.6128  -4.8875  -9.5891   5.6605
#>   -5.5028  -5.2377  -2.9440  -2.6180  32.8425   5.2747  -6.8342  12.1401
#>   -0.5081  -6.8316  -9.1330  -0.1358 -10.8366  -4.8393  -4.7880  -5.4488
#>   -5.5091 -12.5508  -8.3751   0.6286  -0.0728   8.3896  12.7771  -4.4432
#>    6.9994  -4.1967   2.3791  10.3495 -13.5285  -5.4406  -8.2682 -12.8584
#>    4.5179  -8.4492  -4.6975  13.7480  -6.8354   4.7731  20.1120  -3.4452
#>   -3.3951  -8.4500  -9.1310  11.9444 -22.5939  -0.7836  -9.3147  11.4441
#>   -0.8484  -7.1057  12.3151   7.7322  -8.6720  -5.1254  -0.7910   9.3921
#>    0.4421   4.9614   2.2641  -2.8390  -0.9580   2.7702  -7.6955   4.2410
#>   -1.6190 -21.2340  16.0063  -9.9162  12.5261   1.7752  -0.3596   5.2832
#>   -0.0142  10.2848 -10.4606  -8.8678  -3.8597  22.8540   3.4713  -0.4840
#>    2.2803   4.8849  -8.4161   1.5899   6.0032  -9.7212  -0.1825 -14.2181
#>    3.9857  -0.1185   1.2883   4.6189   9.4819  -8.4002  -6.0240   2.5214
#>   -0.7529  19.5940  15.3005  -6.8751  12.2983   5.0802  -5.1074  -8.1556
#>   -1.7678   3.6487   1.0421  11.5745  -7.6581   3.1651 -10.8567  -9.6844
#>   -8.1873   0.6483   5.6380   1.7521  -9.5962  18.8131   8.7751  19.6297
#>   -1.7146   6.4191   7.8329   1.1842   4.5705   7.0575   3.4137 -16.5130
#>   -1.6225 -11.8796  10.3565  -4.4597  10.1547 -15.8071  13.3888   0.7492
#>    5.1109  -7.2423   3.2143   3.4344  -0.9010 -12.2509   6.6927   9.5815
#>    1.1659   9.6458  -1.2745  -5.3798   3.4864  -2.7119  -9.9541   5.3803
#>    6.0919   0.5145  -0.1711  -9.6207  -2.3769 -10.9913  -2.5944   5.8770
#>    0.3253   8.1357  -8.2511  -6.0068  14.4899   6.3091  -3.6282   0.7911
#>    5.0006  -1.0618   3.1083   6.9127  -0.8651 -14.6026  -7.2227 -15.0733
#>   -4.7771   6.8319  12.8300   9.9020   4.1339   2.0625  -2.9212  -3.9819
#>   -8.5579   8.1452  -4.2735   7.6345  15.8145  -6.5311 -10.4794  10.9886
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]