Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.6943 -10.9978  -5.1290   2.1512   8.5171 -10.2894  -0.9582  11.1161
#>    0.3801  -2.7786  -1.9577  18.4473   4.1347   2.6556   5.4092  -1.8762
#>   -1.1221  12.7131   2.8002   4.2606  -1.9475  -6.4263  -8.7638   0.1990
#>   -0.5421   7.1009   3.4933   3.5560   2.6920   5.5756  -3.9643   3.2579
#>   -2.4666   6.8440  11.6558  -4.9910  -0.0696  -3.3286   7.1261  -5.2530
#>   -1.4818   3.4723  -4.8123   2.6319  13.9389  -7.3210  -8.3371   5.9622
#>   -6.2400   4.1671   2.4435   0.4072  -6.1772   5.6128   0.4021 -10.3766
#>   -0.0181  -5.4702   4.2150  -6.2901  -1.6055  12.2291   0.7815  -4.7096
#>   -0.6323  -6.2155  -7.5136  -3.9728  -7.9816 -11.3923   3.0898  11.2577
#>    2.1774   1.9233  -2.1116   0.9434   1.2598  -6.8435  -5.8973  10.3840
#>   -2.3152   9.6617   8.1549  -0.5205   5.2828  -2.3874  -2.2355 -28.4193
#>    3.8850   9.9540   4.1395 -10.4857  -2.5372   3.8763  -2.0852  -8.8911
#>    7.7615   1.6327  -0.6223  -2.9860   2.0155  -4.4935  -6.0561 -23.4585
#>   -2.1236   7.2474   4.4249   3.1680  -3.4762 -16.1750  -9.6007 -18.3559
#>   -3.3402   0.5766  -6.9750   3.3282   7.3521  -3.1926   0.3434   5.7623
#>    4.2901   4.1455   5.3013  -2.6527  -6.5961   7.0399  -8.9388   8.7044
#>    4.1182  -6.2242  -3.5422   8.2539   4.5814   7.3239  -4.7137   1.6172
#>    3.4204  -9.8831   7.0665   6.1551  -7.3386 -12.8807 -12.2144   5.2533
#>    0.7010  -1.3833  -0.7817  -6.1633  -6.0795  25.2649   6.0409 -17.5425
#>    2.4261   3.0930  -4.4735 -12.2193  -9.8971   6.9622  11.8783  -0.7133
#>    1.5973   3.6865   3.1232 -11.9672  -4.4267  21.1312  -0.6718   0.5603
#>   -7.1950  -2.0072   4.4167  -5.0845  -5.3938   8.4179  12.3372  10.0066
#>   -0.6240   3.7722  -4.0336   2.0969 -15.7272  -1.3345  -5.8827  -4.3522
#>   -3.1851  -3.5849  -4.6838  -0.5614  13.7273 -10.7995  -9.1930 -10.1636
#>    4.2198  -2.5085   0.7082  -8.4782   1.0123  -4.8543  -4.9449   6.3214
#>   -1.9310  -1.9797   6.2022  -5.3374 -14.6743 -16.6941   7.7831   7.8185
#>   -9.2593   4.5379  11.4902 -10.6403  19.3052   7.1093  22.4904 -24.3481
#>   -4.7980  -6.7223 -11.8454  -0.6567   5.5463  11.1919  17.8358  -3.2300
#>   -2.9444  11.7642  -6.9590  -7.8556  11.1538  -1.3979   1.2371  -6.6542
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]