Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.4731   3.8765   8.2225   2.7125  14.7183   4.0934 -11.1988 -11.1571
#>   -1.5870  -4.3243  10.5684  -8.8288 -10.5918  16.3526  24.1828   3.3282
#>   -2.1130  -2.3935  -0.4961  -7.3789  14.7191  -7.3337  -1.1690   7.3124
#>    3.8522  -0.2100  -5.1883  16.8012  10.5070  -8.2746  17.3321 -15.5463
#>    1.6860  -3.4782  -3.6669  -0.8998  -5.9221   8.0022  14.7902  -4.3455
#>   -0.6187   2.7607  -0.9922 -15.1709  -5.7528   3.3020   8.3626  12.6313
#>    2.8992  -5.1490   1.5395  11.1359  -2.9926  -0.8155  -0.1483   3.7381
#>   -5.9122  -8.7048  -8.5822   4.6167   4.7561  -5.2771   1.3767  -5.1953
#>   -1.6700   2.8072   2.8957   9.8837 -12.8969  -4.4537 -17.2880  -7.5331
#>    1.6917   1.7218  -4.8800  -2.8023  -4.0624  16.4899  11.9179   8.7013
#>    1.0796  -7.3840  -9.7059   1.1622  -3.6357  19.0743  -3.6835   1.3359
#>    1.7150   0.5735  -6.0632   2.2062  -2.9603   0.8594   4.2432  11.2440
#>   -5.0062  -2.7625  -4.1645  -3.1881   9.9078   4.0923 -17.9519   9.2994
#>   -0.7284   5.5439  -5.7589 -11.7620  -5.6318 -12.8673   9.3886  14.4240
#>   -5.2567  -2.8860   3.8227  17.4813  15.2308  12.1834   2.4403  -2.8418
#>   -1.6279   3.3681   3.5740   3.4814  -1.9374   7.7243 -18.6500 -13.1123
#>   -9.4514  -5.6964  -7.0738  12.4919   9.9957   2.9667   9.9581  -7.7009
#>    2.9157   5.0663   8.2440  13.9585   9.2266 -11.4922 -25.4886  -5.3206
#>   -1.9823  -4.4533  -0.0770   1.8509  -5.9704  -2.4314  -0.2419   4.4249
#>   -2.0706  -3.3026   7.7659   1.3148   9.5339   1.4052   5.1931   2.9510
#>   -3.9552  -3.2571  -6.1464  -7.1704   5.7027  -0.7126   4.8669  -2.5464
#>   -0.3093  -0.6091  -9.3487   3.0323   4.8048   4.9793   7.7388   7.3357
#>   -0.8459   1.8684  -0.1948   9.6375   0.9272 -10.7695   9.6782  -8.1353
#>   -0.0453  -2.3137  -2.3625  15.6322  14.7135 -12.0661 -15.1951  -8.8074
#>    2.6564   0.0022  -3.4828 -24.3115  -1.9875  20.2095  24.4858   7.6600
#>    1.7174   4.3527  -3.7914  -7.3076 -11.5304   8.8435  13.6180  -2.8894
#>   -1.1269  -1.0881   2.0182  -4.6350   3.3187 -13.1533  -3.6574  10.9359
#>    1.2207   8.4922  11.9043   3.5423   5.9217 -25.9484   7.8252  -8.2835
#>    0.6658   0.4731   7.7311  -3.4983  -2.5502   4.5912   9.5963  -7.6997
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]