Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.3841  -1.6480  14.7919   0.7015   7.9894  -0.7795 -10.0584  -8.1209
#>    3.4387   7.3390  -0.0526  -2.8186  -0.2610  -5.1819 -11.3594  18.0839
#>    3.8110  -4.4753  -2.7355  -0.7860   9.1025  19.3252   3.3857   4.1968
#>    5.8966   0.3979  14.6776 -17.1382  -6.1393   1.0477 -14.6533   0.2302
#>    7.5985   2.0664   1.9037  14.8181   0.9565  -7.1625   4.1922  -0.0194
#>   -4.5383  -1.3601  -7.3043   5.6297   0.3255   0.5868   0.9709  11.7426
#>    0.7406   2.0164  -6.9767   9.5803  20.7289   5.8576  11.5107   4.6253
#>   -3.4179  -4.4085  -8.6080   7.4203   6.6419  -7.7847  -0.0428  -3.8476
#>   -6.8137 -10.2102 -12.6369  -9.7697  -1.1413  -6.0231  -3.6013   1.8464
#>   -2.6888   0.9727   2.3770   6.1491  -1.3147  20.0594  -9.5934  -0.7629
#>   -9.3787  -1.5786  -8.6436   5.0011  -7.5743  -4.0205   3.1586   6.2898
#>   -4.8280  -6.2384  -2.1677  -0.8423 -10.6126   3.8987  -1.0601   2.3780
#>    1.5070   1.0678  18.2412  -8.3031 -14.1215   0.3946  -5.7989   0.2242
#>    9.1588  13.0450  -2.1796   0.1627  12.9988  -4.0242  10.6926   7.7251
#>    1.6722   4.5732   4.0071  -8.1771   1.6364   2.6999   8.9674   2.7800
#>    6.5797  10.7889  -2.8672  22.6954   6.9439   6.9677   9.6810  11.2763
#>   -2.8896  -6.1897   4.0433   1.3242   1.1142   7.2812  -4.7890   6.8597
#>    1.8362  -0.2141  -5.8621   4.9214  -4.9851   0.9984  10.3007  -9.5435
#>    4.9176   0.7946 -16.0122   2.3951   4.6073   3.0002  14.6136   4.8664
#>   -1.0937  13.1290   0.9193  -2.5547  -2.9430  -0.6845  -8.6828  -1.8269
#>   -3.8861   2.4778 -10.2184  -1.4099   2.7358  -9.4271  -3.8532   3.5322
#>    1.3458  12.6018  -4.5551  15.6399  11.2256  -7.5292  14.0776  -4.2552
#>   -2.0589  14.0606   4.4167  -0.3071  -2.5300  -9.0268  -7.1638  -3.2108
#>   -6.9480  -5.7546  -6.1820  -2.0374   8.4874   0.6874  12.5008  12.9211
#>   -1.8220   4.0361  -3.2226 -15.7463 -15.1653 -10.9685  -1.0996  -3.5475
#>    2.4436  -1.7647 -20.1564  -4.2842  -5.0181  -8.8958  -0.1498  12.9374
#>    7.0838   7.1681  -1.3484 -11.9666  -9.9269 -17.4642   0.3297 -16.3457
#>    5.8902   9.1700   1.8487  12.4402  12.7816  -1.6469  -3.8646 -24.3626
#>  -12.1299  -8.1948  -9.2063 -18.5146   2.6612  -1.0500  -2.8912   8.9553
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]