Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.9513   2.3013  -2.2428  -3.5291   7.0185  -9.5784 -24.8499 -13.9034
#>    5.5039   1.0452  -2.9146   9.9586   1.7699   5.3183  -0.0148   0.8686
#>   -5.0488  -1.3845   4.1247  12.9997   6.2127  12.2238  12.8604   5.4445
#>   -0.3851 -11.1041  -3.3369   2.0740  -9.2120  -3.7482 -11.4860  12.7668
#>    4.7486   7.4413   1.3538   2.5473  -7.8624   3.8819   7.3091  -5.2531
#>   -0.8174  -1.3008  -5.1279  12.2631 -15.9835   8.3670  -3.7383   5.1108
#>   -2.6870   1.1311  -9.9868   9.3189   5.4522  -5.9236   1.4046   9.4330
#>   -0.1962   1.2755   4.3128   3.0366  -7.1313  -6.0021 -17.8932   3.8783
#>   -0.7471   4.4575   8.2119  -3.1887  12.3711  -3.0705   1.4227 -10.8726
#>    0.9957  -6.5751  -5.8250  -1.0964 -15.0156 -12.5289 -22.5745  -0.9782
#>    2.6441  -4.3119   3.8219   9.5691  -8.4508   0.6930   3.1733   1.0497
#>   -1.1714   3.8656   1.3599 -17.1222   8.7126   8.6363   2.3414  -6.2965
#>   -4.1040  -8.9999  -4.3780   3.0914  -6.2740  -7.8255   8.0831  24.4863
#>   -1.3920   0.5935   1.1146   7.9610   5.0943   8.6134   9.8899   6.1535
#>   -1.9911  14.8864   2.0803   2.4010 -16.0103  -4.6781 -10.1396  -5.8460
#>    5.6964   1.3633   2.3989  -2.6396  -4.2207  11.5630   2.0295  -4.0421
#>    6.9583   0.7215  -1.1646   0.5721  -9.1179 -20.6873  -7.1280  11.2610
#>    0.5225  -1.5730   1.2541   0.5331   6.6026   1.7097   7.4829   2.6297
#>   -2.2655   2.5763   5.9060  -0.4927  -0.8082 -20.2717   3.5135  -0.4924
#>    5.1745   6.9969   0.7166  -2.0968   5.2911  -6.7214   5.5952  16.9880
#>    3.4936   9.0991   4.2432  -6.7836  -7.8147  12.0481   7.1730 -17.6895
#>    1.8236   1.1398  -1.0559   9.9577  -6.6121   5.2728 -16.7422  10.8358
#>   -1.5651  -4.1063   9.9717   4.3816   4.8541  -5.0377 -12.7254 -17.5957
#>   -6.5632   2.5970   4.9909  -5.8259   9.6032  -8.8240   2.2118   3.1999
#>   -0.8761 -10.4677   0.3870   2.8036   8.7853 -12.3970   1.0590  -1.5721
#>    2.8173   1.5636   8.5816  18.5967   8.6646   7.6010   2.7162  -9.0943
#>   -7.2059   1.8472   0.8490  -5.5953 -12.2491  -2.2797  22.3396   8.1859
#>    3.0099   0.3849  10.0426   1.0842  -1.4303   1.9278  -4.3061  -0.7158
#>   -2.4181  -1.2963   7.4685   6.7895   0.1194  -2.4960  19.7261  -5.6724
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]