Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.2495  -0.2606 -12.2639 -17.0907  10.3180   7.9584   1.4664 -14.0022
#>   -2.5195  12.3144   4.9080   1.1604   4.1058  -2.7130  -4.9637   0.1997
#>    5.7257  -1.7812 -15.8917  11.2902   1.6106 -14.1441  12.4136   4.2832
#>    6.0032   6.3549   6.9557   0.0098  -3.4084  -8.4798  -8.7365   5.7563
#>   -3.4972  -7.2365   6.5153  -9.7495  -4.8570 -11.3528   4.6067  -3.4076
#>    1.7876   6.0465  -2.9206  -6.2545  -6.0648  -4.6004  -1.6400   6.6828
#>   -2.3372   7.2080   2.9066   6.2438 -10.5069  11.0101  -3.5214   5.8154
#>   -2.8185   3.0348  -7.9000   1.1905  -8.6505  -1.8362   3.3480  -5.9547
#>   -7.6638   1.8921  -2.6570   1.5969  12.0847  -6.5738   9.1131  11.3694
#>    5.0259  -2.2384 -13.7550  -4.8532  -0.0658  12.4068   3.1467  -3.8407
#>   -0.0541   0.5210   2.1958   9.8524   2.2431   2.7375   4.0363   5.1327
#>   -3.5229  -3.0908   4.0034  10.3187   6.7705 -16.2351  -0.8628  10.0860
#>    1.6678  -1.1670  -5.6025  -7.3798 -13.5995   2.7539  -0.4445  -0.2926
#>   -4.5329  -4.5196  -4.7002   4.4952   6.1060  -5.5428 -12.9214  17.1171
#>    0.4153   6.6180 -10.4284   7.5017  -0.2160 -16.4376   2.2901   6.6125
#>   -2.4526   2.6939   4.5988  -8.5205 -13.1838   3.3370  -1.3666  -4.0097
#>    7.2795   1.3085   8.9319  -3.8162 -15.2252  -7.6737   3.7304  -5.2232
#>    2.4937 -11.5221  -3.7924   4.7708  -3.7473  -2.7503  -6.9877  12.8348
#>   -0.5452  -2.5636   0.9485  -8.6695  11.3579   2.4557  -8.6155  20.2594
#>    1.8440  -2.4776  -6.7151  -1.0816  -0.6525   8.7603  -1.5354   2.7066
#>   -1.6125  -0.0355   0.5367   0.9519   1.9474  -6.8845  -3.2313   1.5454
#>    1.6627   8.3785  -2.9931  -6.3220  -2.9751  -3.7664 -15.4494 -13.3800
#>   -4.3942  -6.6810  -5.0941 -15.5064   2.7118 -16.8134   1.3065  -6.9425
#>   -4.5829  -1.5580  -1.2707  -0.1273   0.7901  -5.2874  -2.8227  -1.2491
#>   -3.4711  -7.5874  -7.3181  10.9144   1.7435 -14.0169  -0.8248 -10.6238
#>    2.7820   2.9940  -2.7928  -0.5128  -2.0322   3.6723   6.1872   6.5101
#>   -0.2115  -4.1447   6.9132   9.0657   9.4754   9.6981 -14.4996  -0.0608
#>    2.9664  -5.4076  -4.8611  -7.4570  -4.9884  -1.0629  -1.7696   8.8946
#>   -0.4160  -4.9115 -12.8091 -12.7695  -0.0021  -3.8515  -9.8569   8.6444
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]