Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.8387  -0.4232   6.4182  12.0635   5.1097  -7.5557  -7.3449   1.9641
#>    4.0775   7.8204   5.2086   4.9110   0.4468  -4.5471  -5.7241   2.7298
#>    0.9697  -4.7948   1.9701 -11.7904   6.9844  -5.2994  -8.9414  -0.4272
#>    2.7219  -2.8507  -9.8149   4.8224   9.4963  -2.8199   2.5106  -2.7543
#>    3.6816   2.2898 -12.4248 -15.5545  -7.2955  -2.8228   5.1010  -3.2680
#>   -0.6888  -6.1766  -2.9543  -2.9418  -7.7383  -2.1010  -6.4469  -9.3888
#>    4.5773  -4.9552   6.4707   3.3034   4.8926   1.9464  -0.1015  -2.0312
#>    1.5054  -2.2201   0.4528   4.7132   4.1934 -13.0134   1.6217   1.9625
#>    0.3502   6.1899   2.0187   3.7795  -3.5631   0.4334 -16.4787  -1.2164
#>    4.3739   1.1778  -2.1286   7.8198  -3.8965  -2.9405  -3.5529  -0.2562
#>    6.5704   9.9048 -10.0359   2.1084  -1.1238 -12.1627 -12.6528   1.1302
#>   -5.7436  -6.7936  14.7420   5.0860   2.0330  14.2526  -4.1570  -2.8185
#>   -0.4281  -5.0399   0.3582   7.1698  -4.4327   9.7260   5.8410  -1.6119
#>   -1.4509  -7.4750   2.6865   6.8045  -2.2284  13.0663  17.0327  -9.1340
#>    4.9087 -12.0510   0.8855   8.2191  -1.8712  -8.2972  -0.3709 -12.6974
#>    0.7013   1.1327   1.5076 -15.8341   7.8765  -9.9662  -0.2475  -1.2303
#>    4.1809   5.1597  -4.2921  -8.7612   4.0096   5.8517  -0.3606  -8.0625
#>    0.9000   5.9561   3.0011   0.6288  10.7118  -1.1947  -5.6200  -4.4821
#>   -6.3000  -1.4697   4.0011 -12.5426  -6.7954   3.0030 -13.1807 -11.5293
#>   -3.4724 -12.9334   1.6021   3.2470   3.1076  13.6691  19.6328  -5.0149
#>    7.6034   3.2222  -1.5798   3.1423  -4.2024  -0.4271  -2.5390  -8.4036
#>   -4.1443   2.4504  -9.3807   2.3166   7.0624   7.9100  -7.8894  -5.6617
#>   -7.5814  -1.1609  -7.6857  -0.9912  -7.6945 -12.0192   0.6484  -0.3359
#>   -3.5725  -0.4694 -10.2874  -5.3361  19.6918   7.1659   3.6238  -5.0994
#>    2.9922  -3.7347   6.6455  -7.9693  -9.2763   1.0766  -0.8153   0.1435
#>    8.6997  -6.2224 -12.1242  -1.5732   5.6334  -3.6370   2.5678  -1.2040
#>   -1.0716  -7.5925  -3.8443   2.6544  11.4202   1.5301  -0.8370 -18.6548
#>    4.4266   3.3392  -2.0882  -6.9118   4.3474  -1.8681  -0.0403   9.0745
#>    2.7090   8.7068   2.7224   8.3626  -0.6677   3.3243   5.7600  -7.9481
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]