Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.5041  -3.5002  -3.7107   5.2788  -4.2390   8.5434   4.6750  11.7616
#>    0.6149  -3.6565  -4.7664   6.7826   3.7324  -2.4864   2.4645  -9.0225
#>    0.8424  -0.3674  13.5140  -4.8287   3.0835   1.4393   5.7434  -3.8169
#>   -1.5557   9.0362   5.8932   6.3136   4.1730  26.6048 -17.2025  -8.4015
#>    2.0650   2.7499   4.0992   5.2711   9.7082  -3.8163 -14.5422 -11.2385
#>   -0.2053  -6.1799   2.5997  -1.5469 -25.0018  -3.8108  -1.5562  15.9698
#>    2.8733  11.2445  -7.2614   0.4446  -7.3643  -6.3233  -1.4275 -15.2960
#>    9.3191   0.6891  -9.5947   2.5323  -6.2636   3.0385  -5.2314  12.9090
#>    2.1628   0.4822  -7.7300   9.2803  -7.7407   3.2303  -4.1930  -4.9030
#>   -5.9003 -12.1276   3.0118  -5.8857  18.7737 -12.5288  -2.8115  -0.2312
#>    0.5518  -1.7369   9.3000   0.9632 -10.8654   5.8339 -11.5655 -17.4844
#>   -3.0082  -6.1118 -14.5426   4.7904 -13.8884  -9.6703  21.4249  10.0689
#>   -2.1047   1.4233   3.6173  -2.9593  -2.2854  13.0185  22.2353  -6.8339
#>    3.2763   4.6213  -0.8961  -7.4057  -8.4863  -6.9612  -9.8720  -6.4665
#>    5.4683   2.7604  -0.1172   0.7614   9.5472  -9.3949  15.2087   4.9862
#>   -2.4749  -3.3320  -4.0411   1.5837  14.4594  -4.0278   2.4856   3.3730
#>   -0.5834  -3.1912   6.7686  -1.2936   3.2552 -11.5498  -7.5710   6.9197
#>    2.4156  -1.9325  -8.1143   9.7194   5.1750  -1.5398  17.6628  -2.0831
#>   -2.3108   7.1940  -6.7130 -10.8878   1.4480  -5.5551  -4.1614  -9.5437
#>    5.0038   5.1786  -6.1255   4.4437   2.2757   2.8566   2.8580  11.2384
#>    1.0449  -3.8385   5.5213   8.1822   4.4289   1.9017   4.2447  15.0696
#>    1.7886  -0.0262  -0.8159   2.7761 -12.6435   8.7623  -2.3110   5.9581
#>    0.5399   2.5752  14.9910   7.8349  -5.3652   7.8347   2.9462  -9.4957
#>   -1.4413   4.5547  -8.7895   9.9336  -6.4407  -7.1059 -15.1875   0.0270
#>   -3.9968   7.9517  -7.2836   3.8712  -4.7450  -6.0372   1.8028   6.1005
#>    4.2749  -5.5423  -4.8383   9.9014  -1.6981   2.8573   8.0989 -12.7256
#>    4.3394  -5.0038  -0.5341   3.9705  15.4474  -1.8034   2.1544  10.9410
#>   -1.5900  -1.5831  -4.7079 -11.2273 -10.7603   4.4207  -5.6564   6.2413
#>    2.3107  11.0584   0.7939   3.7578  13.7170  14.8473  -0.5479 -11.8654
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]