Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.1749   0.4227   7.1118  12.0020   7.2109   0.0364   3.3119  -7.4843
#>    0.8703  -5.4373 -20.2264  -4.4400  -9.7750  -3.1685  23.1825   1.6904
#>    3.4632   3.8288   7.0512   5.3602  10.3797  -8.9597 -10.8714  -5.5104
#>    3.8972   1.4619  -7.3029   5.6908   7.3144  -9.9437   3.6504 -11.6614
#>   -1.7972   5.3135   4.0258  12.1217   8.0168 -19.7540  -3.4393  12.8293
#>    3.9853 -11.1517  -1.5517  -0.5036   0.8791  11.9280   8.1400   1.7993
#>    0.7123   1.4117   0.5595   2.2990  -3.8834  22.7326   4.6402 -17.1024
#>    1.4612   1.3120   1.2181  -9.8951 -10.5543  -0.1949   5.7005  13.5972
#>   -1.7997   6.7036 -15.0104  -1.0406   7.0556   7.0636   0.2549  -0.4601
#>    3.2517   4.4534   4.6419  -4.1084  16.0512   6.5165  -6.8026  -1.2657
#>   -2.7923  -0.1344  -5.1198  -4.5210   0.6464   6.6175   4.2729   6.4027
#>   -1.7921  -5.9831  -2.2252  17.9070   0.4135  -0.4538  19.3850  -6.3850
#>   -2.5328   3.1228   4.0845 -12.8365  -5.9409   0.4080 -15.3684  -2.2581
#>    3.1998  -2.9186  -2.7877  -9.9578  -4.8281   1.1315  -5.0721  -2.5338
#>   -2.4352  -5.1534  -3.6513   0.7693   4.9778  -0.3628   5.1657  -0.3403
#>    5.1228   3.2690  -1.0407   4.9541 -14.7330   5.4304  -9.0368 -21.6179
#>   -3.0341 -10.2939  -9.9268   3.6777   3.4548  13.9607  15.8807  -5.2298
#>   -0.4666 -10.6917   0.0436  -6.7592   2.2505   6.9622  -3.2386 -13.0587
#>   -7.2650   4.5958  -1.9393   8.8183  10.7259  20.0067   3.5051  -4.9370
#>    0.6985   2.2462  13.1570   6.0991 -21.9163  -2.4901  -7.9133  -4.0806
#>    2.8646  -7.2457  -2.2613  -0.5152 -11.9879  10.7566   6.1536   1.6236
#>   -0.8088  -4.1866  -0.8687   9.5549  -3.7326 -13.4855 -21.0415  -3.5882
#>   -5.4249  -7.0632  -5.6878  13.6657   7.6758   7.0155   8.9586  -1.4923
#>   -4.0998  -7.2423  11.3709  16.1749  -7.5660  -6.5412  -0.1843  -0.0098
#>    1.9998   1.5740   0.6116   3.2075  12.1019  -6.9942  -5.9076  12.2184
#>    0.3853  -0.7995   6.2496   1.9950   3.9805  -3.9180   0.7548 -11.3195
#>    4.0748   0.9931  -2.0987  -8.8949  14.7991  -4.3501   9.0467   3.5010
#>    0.3417   0.5464   6.4674   0.0696  -1.1304  -5.8489  -4.5582  -4.7075
#>   -5.7862  -5.4786  -1.1976  -5.2212   8.4610  13.8050   6.8904  -1.4175
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]