Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -5.2668  -1.8317   4.1291  -0.6012  -2.4646   8.2833  12.9509  -0.4013
#>   -3.5413   7.3400  11.6791  -5.4318 -12.1386   3.7249   1.7071 -18.5994
#>   -1.0436   8.2000  -3.4043  -0.7808   3.0552   2.0143  10.4898   3.1205
#>   -3.8170  -1.5770  -7.2323  -7.1191 -10.3916   4.9890  -9.7526  -5.8625
#>   -0.1379   9.8747   1.9460  -7.5252   1.1071   2.9822 -17.9717  -2.0371
#>    0.4386  -2.0179  -1.4356  -2.0929 -13.9245 -16.3379  -1.6855  -9.6339
#>   -3.3015  -6.2208   3.6997   3.3138  -0.7798  -3.7128  -2.5994  10.9692
#>    4.3457  -2.6494   4.5258  -1.9266 -11.8598  -8.8285   1.4577  -8.4091
#>    5.4584   0.7350  -3.7826  10.3698  -2.9493   8.9109  -1.9176  17.6626
#>    0.0565   1.7213 -11.7932  -5.9302   2.6835  -9.0777   4.6582  -0.4141
#>   -4.7854  -0.6254  -2.1750  -6.2648   0.0555   2.8375   9.3931  -7.4711
#>    5.8272  -6.2862  -0.5341  -1.3463  10.8596 -12.9811   3.3870  10.1154
#>   -2.4331  -9.9516  -4.2067  10.0938  10.1617 -31.6054  -3.9211  24.3192
#>   -1.3134   1.6995   1.1477  -2.3798 -14.2363 -21.9547  -6.5992   2.2153
#>    0.7988   3.0862   3.4568  -6.9066  -1.3726  -2.8589   2.6746  -6.1005
#>    0.1842   7.3305  15.7106  -6.1831 -10.8783   1.9882   9.0140   1.4316
#>    4.7721 -15.4879   0.3949  -1.2214  -7.0618  -7.5915  -6.5987  -5.1631
#>    0.1592   3.5073  -1.8558   1.5173  -2.9913  -9.0528  -5.7528  -2.8652
#>   -0.0690   3.2651 -11.6396  -4.8850  -5.3705   3.2849   0.4004 -13.4383
#>    5.0900  -1.8387  -4.7794   5.4331  10.7239  -4.6893  -4.0167  11.5067
#>   -5.3388   0.2179   7.8621 -11.6612  -6.1536   1.8431   6.3677 -11.4275
#>   -5.4878  -3.7875  -2.6788   9.8348   3.0959   1.1328  -9.7551   8.5331
#>   -0.0843 -10.1828  -7.4964   1.3989   4.9373 -10.0752  -3.0322 -18.7779
#>   -5.2475  -0.4882  -8.6748   2.9233  -0.1050  -2.2566  -0.9708  -1.2344
#>    5.6475  -1.0109   1.9554  -0.5339  -3.4041   0.9602  -6.9846  -9.3388
#>    2.3337   5.0083   8.4899   0.8598   3.8809  -2.1600  13.5354 -11.7252
#>    3.2964   4.0928  -6.4441 -15.5577   5.3379  -5.5075  -2.4785   3.3619
#>    4.8735  -9.9215   3.3732  -4.8606  -7.5492  11.7787  -3.9373  -4.3273
#>    2.1339   4.4550   3.5274   0.4730  -0.9996   4.7639  12.7515 -13.0802
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]