Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -6.0387e+00 -2.9146e-01  5.0521e+00 -3.4515e+00 -6.0757e+00  7.9507e-04
#>  -5.9992e+00 -6.1146e+00  9.4210e+00 -1.0199e+01 -1.0122e+00  1.2715e+01
#>   4.4189e+00 -3.6982e+00 -5.8268e+00  9.7544e-01  6.9349e+00  1.9145e+01
#>   1.1279e+00 -1.2149e+01  1.1968e+00  4.5374e+00 -1.1873e+01 -1.5569e+01
#>   5.3319e+00 -1.4959e+00 -1.1698e+01  5.5405e+00 -3.2498e+00  9.9627e+00
#>   4.9664e-01 -3.2854e+00  5.6530e-01 -1.6260e+01  1.3261e+01 -1.4995e+01
#>   2.7976e+00  1.8954e-01 -1.8086e+00  7.5574e+00  6.7528e+00 -1.6535e+00
#>  -6.6192e+00 -2.3063e+00  4.4016e+00 -2.0452e-02 -6.2862e+00  1.2278e+01
#>  -3.1331e-01 -4.0775e+00  7.6898e-01  1.7764e+01 -9.0590e+00 -1.3214e+01
#>  -9.3346e+00  6.5458e+00 -5.9188e+00 -1.5235e+00 -2.7916e+00  8.1813e+00
#>  -2.6748e+00  6.8343e+00  2.5809e+00 -5.2077e+00  3.7665e+00  2.0089e+00
#>  -2.2076e+00 -4.5626e+00  1.8755e+00  2.1854e+00 -9.8296e+00  1.0504e+01
#>  -3.1852e+00 -5.7878e+00 -4.9948e+00  1.4890e+01 -2.2559e+01 -2.8491e+00
#>  -6.4482e-01  2.2182e+00  5.3812e+00  2.4381e+00 -9.9336e+00 -3.1839e+00
#>  -7.7600e+00 -3.6300e-01  9.3702e-01 -3.5311e+00  1.5673e+00  2.1318e+00
#>   2.3813e-02 -1.8230e+00 -7.2346e+00  5.2343e-01  1.7626e+01 -6.2312e-01
#>  -1.1970e+00  1.0495e+00 -8.2544e+00 -7.9753e+00  3.4747e+00 -1.2783e+01
#>   4.3123e-01  5.2179e+00 -1.2133e+01  6.2660e+00 -7.7262e+00  4.0990e+00
#>  -4.7728e+00 -3.4791e+00  7.2419e+00 -1.0408e+01 -1.8996e-01  1.0338e+01
#>   1.0426e+00 -8.9125e-01 -1.0459e+01  3.2134e+00  4.7735e+00  1.6278e+00
#>  -8.5254e+00  3.7046e+00 -1.8980e+00  3.8967e+00  3.6429e+00  5.5321e+00
#>  -6.9400e-01  2.8877e+00  5.1449e+00 -1.4690e+00 -2.7853e+00 -5.1867e+00
#>   7.9984e+00  1.0710e+00 -1.0427e+01  3.0328e+00 -1.5604e+01  1.9222e+00
#>   2.2159e+00 -3.7553e+00 -1.3656e+01  2.2782e+01 -1.9312e+00  6.0407e+00
#>   3.7142e+00  9.0469e-01 -7.6449e+00 -6.7523e+00 -5.4103e+00  9.8024e-01
#>  -3.7975e+00 -1.5329e+00 -6.2346e+00  5.1276e+00 -8.6008e+00  9.9587e+00
#>  -1.1006e+01  8.2813e+00 -5.2642e+00 -1.0448e+01  5.2040e+00 -7.3103e+00
#>  -3.3648e+00  6.5036e-01  5.7394e-01  1.9380e+00  7.5108e+00  7.6064e+00
#>  -2.3359e+00 -2.3514e-01 -1.9962e+00  2.1281e+00  8.8278e+00  7.7421e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]