Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.2721  -3.4198   5.2768   5.2703  -3.3275  -8.1115  15.9925   4.0437
#>   -1.8249  -2.2227   5.8427  16.6108   2.9308  -2.9900   5.8923   6.2257
#>   -0.4552  -1.4893   8.7362  -5.0451  -3.6302   1.8941 -13.2540 -26.7684
#>    1.1562  -0.5850  16.8523   5.4491   5.9403 -12.1012   6.5335   3.7782
#>    1.1837  -3.0845   4.1555  -4.4026   8.6744  -5.1869   7.4616 -15.1954
#>    0.3460   5.9465   6.5456   3.8366   3.1326   8.6348 -12.8678 -22.0440
#>    3.6553   7.3526   1.3759 -17.0522 -11.1613 -15.5233 -20.1521  -1.3245
#>    0.8236  -1.2341  -4.7172 -13.1772  -4.6038   0.0632  -3.1765  -1.5379
#>    8.1115   5.5612  -7.1279  -1.6772  -0.8339  -0.9790   3.2246 -12.0750
#>    0.8969   2.6957  -4.5025  -0.5372   0.0051 -13.8251  12.1838 -12.8477
#>   -1.0703  -4.1722   7.6155   0.4441 -11.4229   4.6001  15.2914  22.2445
#>    8.1084   2.0643  -9.3213  13.1703   9.8371  -6.4055   7.3230  11.9915
#>   -6.2901   4.7002   0.6888   6.3952 -10.8812   4.1801 -11.6051 -15.8973
#>    1.9574  10.0748  11.8300  -3.4655  -4.7836  -5.3502  15.8427  15.6749
#>    2.0804  -6.6037  -2.2404   6.0839 -20.6944   1.2541  13.2789   0.7905
#>    8.4460   7.7297   4.2576   0.4983  -4.8400  -5.8682  -8.7012   9.9860
#>   -6.9826   2.9510  -2.4570  -1.7082  -0.8502   2.8057  -3.2919   2.2968
#>   -6.3654   0.0022   0.7703  -6.5018  -3.0769  12.6562  -6.1853  -9.6171
#>    1.3520  -4.2945   1.3045  -1.2869   7.6168  21.0480  -1.9970  -2.4627
#>   -6.7955   0.5747  -2.1618   6.0766  19.5697   7.4547  -3.2555  23.6683
#>    2.1934   7.2093  -6.9916  -0.5213  13.6782   1.7051   1.0628  -5.5219
#>    2.9324   3.4748   5.9206  -3.6897  -2.9834  -2.3291   9.0634  17.6877
#>   -2.7518  -8.9060  -2.0299 -23.1378 -25.4429 -16.9659   9.4131 -12.7891
#>   -3.6787  -9.9271  -4.0338  -6.7167  13.7053  13.1748  10.9977  -3.1643
#>   -1.7493  13.8509  -9.8119  -5.2957  17.5401   1.8333 -10.3856  13.9012
#>    3.9964   0.3554  -1.4905  -6.2301  10.5557  11.3684  -5.1243   1.1337
#>   -4.5724   2.4417 -10.1535  -8.4239 -23.5330  -0.3234   6.8189   0.4392
#>   -5.5346   6.8105   6.7906   3.1553  -0.8488  13.3237 -10.6394 -11.0128
#>    0.9422  -2.5756  -7.7518  11.7405  11.3872 -11.4613 -24.5801  11.9540
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]