Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  10.2355  -0.8804  -2.8219   2.0782  -1.6218  -8.7457 -10.9500   6.1533
#>  -12.3512  -6.7147  -4.5731 -26.6255   7.0583 -15.7699 -14.5134   5.7863
#>   -8.9900  -2.4053   6.8586 -11.3385  -7.0599   2.5049   3.3010   2.0990
#>    0.1012   6.0849   0.4468   9.4507  -1.3610  -6.7855   2.1841   7.5927
#>    1.9025   3.3778  -7.0083  13.0966   8.8447   9.5793   0.3946   0.5356
#>    1.8886  -1.4720  18.3050   0.0344 -23.8855   8.4348   8.8701   6.0999
#>   -4.7908  -1.3217 -13.4658  -8.8647  -6.1847   0.0602   4.5759   3.0649
#>   -5.1026   3.2726  10.8240  -3.8436  -6.7145 -10.2475  -0.2592  14.7788
#>   -8.2802  -2.1460  -4.3062 -14.0367  -7.7575 -14.1985 -18.1790 -16.3615
#>   -3.5061   4.3300   3.2745  -3.5064  -8.9353  -1.4079   5.2913   1.3493
#>    4.1444   3.7632  22.4203   1.1310   9.0901   8.7382  -6.6166   3.9081
#>   -2.9998 -16.9679 -13.6180  -5.1544   4.1194 -13.5410 -15.0082  -2.9786
#>   -0.5287  -0.4992   4.1011  10.6276  -4.9429  13.8889   6.0311   4.8996
#>    1.1539   0.6948  -4.1329   5.1528  21.4415   1.5378 -14.4336  -3.5669
#>    0.5916   8.8692  -7.3973  -3.2780  -4.1024   7.8517  10.1479   2.1933
#>    0.9083  -6.7905  -2.3144   7.7140   0.2851   9.3610  -4.3375  -3.4518
#>   10.4454  -1.1895   5.8242   4.9367   2.0330   7.4862   9.4862  12.9778
#>    2.2446   2.5625  -1.3983  -9.6085 -13.6675   1.9800  -8.3648  -5.1309
#>   -7.2108   1.8471  14.5990  -3.4068  -7.9009  11.7675  -1.5431   1.6747
#>  -11.0502 -11.8636   2.5344 -14.5597  -2.5933   2.4875   0.1092 -12.2142
#>    1.6047  -2.9818  -4.1089  -2.6615  -7.7452  -6.7504 -22.9932   1.0196
#>   -3.8662  -4.8253   5.8069   6.6950   1.5051   8.5833 -21.6092   2.1116
#>   -4.8659   0.4868   4.4508 -16.1968   1.0940 -11.7425   3.8984  -3.7725
#>   -2.2653   5.3564 -11.6340  -4.6852  -6.1798  -9.6863  16.4693  -3.4198
#>    2.3858  -1.2630  13.8133  -7.0434 -16.7904 -10.5720  -6.8476  -5.7484
#>    3.8919  -1.6161  -1.2256  -0.7735   4.8475   6.0544  13.3500   5.3432
#>   -1.5104  -5.7800   2.5119 -11.6065  -1.1289   6.4375 -16.0118  -5.4732
#>    6.3943   2.3815   6.5277   5.1839   9.8305  -3.0142  -2.3789 -11.9744
#>   -1.0861   1.7431   5.2531   0.2286   6.7437  10.2135   9.0868  -2.8829
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]