Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.4297  -8.4326  -3.5491   2.8363 -10.1400  16.3035  10.6111   7.3374
#>    1.9996   0.7233  -5.0346  -3.8125  -6.4300  10.2876  -4.0279   7.7084
#>    0.5831  -4.3465  -6.0386   4.4836   5.9045 -14.5409  -9.3000   1.1273
#>   -2.9386   4.1621   3.0330 -10.7074   4.4601 -20.7333  -2.3875  -7.7129
#>    3.4822  -3.8150  -2.1031  -3.7253   5.6452   3.2656  -0.2575 -13.8666
#>    6.0293   4.5451  -7.9777   2.5771 -16.7307 -14.0736  -3.8956  -0.5045
#>    1.2001   3.9398  -5.8153  -5.2303  -2.7601   5.8232   4.6911  -7.6124
#>    0.6239  -5.8748  -7.7602  -5.3697  16.4393   0.2864   8.9267 -10.6360
#>   -1.4722  -2.0904   7.7300   4.4306   3.8137  -4.5345   1.0108  -1.5423
#>   -4.7861   1.2469  -2.5956  -6.5660  13.0057   1.4045   3.6812  -3.6755
#>    1.9536   2.4694 -10.6432   1.0399   1.5805   5.0476  -1.8424  11.4907
#>    1.2809   8.4944 -12.4259  -2.1177  -7.0219  -1.9168   5.7135  -1.7295
#>   -0.3942  -0.1763   4.6878   0.9012  14.5979  -6.6922  -3.4672   5.0600
#>   -0.8423  -0.3490   2.2759  12.9629 -12.0749   1.6968  -6.6299  12.2998
#>   -0.8728  -5.4835  -0.1885  -1.3899  -6.2499   4.4132  -5.4307   9.9495
#>   -1.2665  -2.2229   9.6993   0.4512  -4.9480  -6.4565  -6.7521  -3.3942
#>   -0.6676   4.1457   3.4701  -0.4566  -0.5821  -4.4669 -13.6131  -8.7767
#>    6.2853   4.7549  -4.6195   4.8898   7.8393  -6.0062   2.0231   0.4886
#>   -0.1896  -4.4392 -10.9389   2.3837  -5.0042  18.5621  -2.7869  16.0712
#>    7.7474  -4.6128   9.1308   2.7983   8.7379  10.2900  11.5431  -6.5187
#>    0.5667   2.0691   0.6790   0.6172   7.2980 -15.8614  -0.1239 -18.7530
#>    5.6108  -7.6161   8.5916  10.5220  -2.2923 -12.9235   0.3514   1.8670
#>    3.3483  -6.5530  -5.6319   9.3289  -9.5489  12.4105   2.0325  -0.2396
#>    4.7278   5.8927   2.4867   3.4496   7.5138  -8.7773  -5.6677  -0.3573
#>   -1.9425   1.0605  -5.2804  -2.5611  -4.5296   4.6617   8.6738  -2.9357
#>   -0.4734  -4.7419   1.0269  11.2558 -15.9964   0.0914   4.6969 -11.6441
#>    1.3718  -1.8513  -6.6262   7.2395   9.8467   1.0868 -15.3125   5.0592
#>    3.1130  -5.5298   5.4217  -6.6562  -0.0749  -4.8429   6.5599   0.4420
#>    0.1453  -1.5061 -10.2270   0.2346  -5.5203 -17.0299  -6.9567 -15.9207
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]