Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  8.7077e+00 -9.2227e-01 -8.6119e+00  3.8889e+00 -6.3060e+00 -3.8728e+00
#>   1.8233e+00 -2.3361e+00 -9.3346e+00 -4.9568e-01  4.0819e+00  1.2496e+01
#>   8.4293e+00  7.1007e+00 -5.7668e+00 -1.0264e+00  7.2157e+00 -1.0999e-01
#>   1.5981e+00  5.4237e+00 -1.2860e+01 -1.1724e+01  5.2653e+00  1.6635e+00
#>  -1.1065e+00  2.6060e+00  1.9299e+00  1.9561e+00 -1.3482e+00  3.0959e+00
#>   2.5579e+00  3.7822e+00 -6.9048e+00  9.2686e+00 -6.3666e+00  1.2102e+01
#>  -7.3794e+00  3.6009e-01  3.4538e+00  1.0421e+01 -1.6153e+01  7.7285e+00
#>   7.1277e-01  4.7353e+00  9.8236e+00 -2.7900e+00  4.2212e+00 -3.5540e+00
#>  -1.0986e+00  1.5976e+00  3.8099e-01  2.3559e+00  7.6269e+00  5.7011e+00
#>   1.0850e+00 -1.6235e+00  3.8809e-02 -2.0356e+00  9.1069e+00 -3.8679e+00
#>   3.0983e+00 -1.5713e+01 -4.4033e+00  6.2236e+00  1.1097e+01  6.3632e+00
#>   3.9919e+00 -5.4779e+00 -1.0699e+01 -4.0830e+00 -4.7938e+00 -8.0010e+00
#>  -1.2254e+00  6.7389e+00 -6.9358e-01  1.5311e+00 -3.4889e+00 -3.4930e+00
#>   4.3958e+00  8.5936e+00 -3.6723e-01  6.3811e+00 -9.3680e+00 -9.1447e-01
#>   2.1327e+00 -1.0735e+01 -7.5921e+00 -1.0197e+01 -2.1811e+00 -8.8906e+00
#>  -1.4581e+00  1.2470e+00  1.8829e+00 -4.0160e-02 -6.9690e+00 -9.9223e+00
#>   1.5186e-01  5.5656e+00  7.7329e+00 -9.2243e+00  2.3871e-01 -5.2758e+00
#>  -1.1178e+01 -3.8008e+00 -1.2281e+01 -7.3901e+00  2.0565e+01 -9.8619e+00
#>  -1.1190e+00  4.5475e+00  5.0468e+00  3.3386e+00  6.1725e+00  1.7424e+01
#>   2.3825e+00 -6.4268e+00  2.2578e+00 -5.5909e-01 -9.8564e-01 -5.3610e+00
#>   1.7426e+00 -8.1060e+00 -1.8271e+01 -2.0841e+01  1.8399e+00  1.1165e+01
#>   8.9400e-01  3.3499e+00  6.3694e+00 -1.1418e+01 -2.3793e+00 -1.0863e+01
#>   3.5550e+00  7.7069e+00  8.7315e+00  6.7881e+00  1.4667e+01  4.3719e-01
#>  -9.8039e-01  1.2773e+01 -4.1325e-01 -1.1850e+01 -1.1958e+01 -1.4193e+01
#>  -8.1063e+00 -7.2632e+00 -4.3897e+00  1.6251e+00 -8.8857e+00 -3.3409e+00
#>  -2.3085e-01  8.3850e+00  9.8876e-01 -3.0562e+00 -4.7517e+00 -4.2008e+00
#>   8.5259e+00 -1.8074e+00  1.1095e+01  5.7336e+00  1.3445e+01  1.8439e+01
#>  -4.3691e+00 -5.1397e+00  6.1225e+00  1.8792e+01 -9.0471e+00 -3.8916e+00
#>  -7.3966e-02 -2.8531e+00  7.5487e+00  8.6243e+00  1.4485e+01  1.6561e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]