Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.1354  -6.4509  11.9339   3.6335  10.4627   3.0479  -6.2967   5.4741
#>    3.6351   7.5755  -0.1116  -3.1910  -4.6949 -12.1842  -0.3792  -7.0186
#>    7.3572  11.9695   0.7709   0.0844  11.7651  -1.7083 -10.8820   1.6327
#>    2.0608   8.7098   1.7127  -2.3917  18.5255  -4.2179  -8.4208  -3.9780
#>   -8.6462  -9.3310   6.7517  -2.2810 -30.3501 -18.6527   4.4170   1.7945
#>    8.9266  10.5263  10.9597 -17.3916  -7.3251  -2.4004  -1.5277   3.9908
#>   -1.9600   1.1091   9.8785  -0.2046 -13.1723  -1.2965  -2.7618 -10.4532
#>    2.1791   0.3149  -0.4514  -7.2782 -11.2823   7.9705  -0.3488   3.8664
#>   -7.0747  10.4319  -0.2562   8.9613  -1.3923   7.7114  -1.9964  -3.3544
#>    6.5591   2.5620   8.7340   2.1693   2.7984  -5.1666  -0.1489  -4.1032
#>   -5.6935  -9.7082 -20.1218  -6.2220   2.3419  -4.8840  -8.2903  -8.4664
#>   -0.3110  -0.2066  10.6637  -6.0859 -13.1089   0.2286  -0.1388   9.0878
#>    7.3038   0.8211  -5.4876  -6.9590 -17.3400  -3.0886   7.2908   0.8545
#>   -4.2778   4.7120  -6.8512   2.6053   4.5279   3.0507 -14.0170  -2.0648
#>    2.0220 -16.4871 -20.3636   2.4171   8.3917  -8.2009 -14.9715  -2.8813
#>   -3.5552   5.1126  -7.0629  -0.1762  -2.9632  22.1081   0.9665 -11.1933
#>   -4.8639   3.7862   7.1010  -4.0702   1.2361   8.6701   4.6009  13.1383
#>    2.6356  -6.0301  -5.8626   8.9551  -5.5431   8.6916   7.8152  -2.2973
#>    2.3156   9.5163  -7.8860  14.8711  -0.6482  -6.4438  10.7748 -10.0511
#>   -3.9210   1.6805  -2.3526  -4.1793  14.6970  15.6459   0.2950 -20.3189
#>   -5.1242  -0.2931   2.2740   8.3453  -9.2923  -5.9872  -3.0189  -2.4312
#>   -4.4542  -1.0150 -15.5950 -20.8395   2.4091 -10.0613  -4.0997 -18.7938
#>    0.8841 -11.3296  11.9088   2.1987 -24.0058   3.8613  12.2137   4.7085
#>    4.0157  14.9103   6.4363   5.5108   9.4516   5.8887  16.0960  -3.2763
#>   -4.0831  -1.7765  -8.6103   3.8849   1.4219  -3.9209   3.4961  -0.8402
#>    1.4439   6.5642   3.0173 -11.4977   8.1838   1.8115   2.1024 -17.4182
#>   -0.7435  -7.3470 -13.4172   3.4689 -15.6388   1.7653  -6.0858  -4.5273
#>    2.8378  -7.4041 -10.4009   1.2184 -15.0786  -9.0820  -6.3717   1.5123
#>   -2.0109 -13.5822  10.0636   6.9495   0.5328  12.1477  10.8075   9.8340
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]