Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -3.1714e+00  1.1062e-01 -6.7955e+00  2.9056e+00 -9.8881e+00 -4.6030e-01
#>  -4.4099e-01 -8.2616e+00 -1.0330e+01 -5.6497e+00 -1.4280e+01 -6.5590e+00
#>   3.7544e+00 -2.9258e+00 -6.7598e+00 -1.1615e+01 -1.4603e+01 -1.4882e+01
#>  -5.2511e-01  6.6658e+00 -7.0567e+00 -9.4705e+00 -1.7706e+00  3.5249e+00
#>   1.9710e-01 -2.4995e+00 -2.2341e+00 -1.3710e+01 -1.8914e+00 -5.1356e+00
#>   1.9047e+00 -1.4756e+00  7.7721e+00  9.1372e-01 -2.1540e+01  4.0249e+00
#>   3.3316e+00  1.5004e+00  3.7345e+00  1.4011e+00 -1.2512e+01 -8.1744e+00
#>  -6.7904e-01 -6.0267e+00  1.9722e+01 -8.7920e+00  1.0926e+01 -7.2644e+00
#>  -5.4948e+00  3.7591e+00 -9.3395e-01  1.2491e+01  6.0419e+00  1.6641e+01
#>   4.0539e+00  5.3998e+00  2.8916e-01  6.5487e+00 -2.6070e+00 -2.9460e+00
#>   2.3271e+00 -4.2677e+00  2.8139e+00  7.2740e+00 -8.1247e+00 -1.0798e+00
#>   3.4054e+00 -6.5511e+00  5.2185e+00 -5.5927e+00  1.7701e+01 -9.4807e-01
#>   1.6934e+00  6.2735e+00 -2.7209e+00 -7.8939e-01  5.6247e-01  5.1325e+00
#>   1.1680e+00 -6.5091e+00 -6.3331e-01 -1.7060e-01  4.4019e+00  1.3833e+00
#>   1.6583e+00  7.4128e+00 -1.3918e+00  2.4850e+00  1.0738e+01  1.8250e+01
#>  -4.0462e+00 -3.8598e+00  1.1394e+01 -1.4305e+00  7.1592e+00  1.4612e+00
#>  -4.0910e+00  1.6125e+00  1.0468e+01 -1.1892e+00  1.1369e+01  3.1122e+00
#>  -1.1169e+01 -5.3737e+00  8.4064e+00 -9.6351e+00  7.4874e+00 -3.5551e+00
#>  -3.6654e+00 -1.8686e+00  1.0466e+00  6.3591e+00  1.9117e+01 -1.3530e+01
#>  -4.1374e+00 -2.1420e+00 -9.0342e+00 -6.0437e+00 -1.1078e+01 -3.9243e-01
#>  -2.7481e+00 -2.1509e+00 -6.7437e+00  9.7476e+00  1.7671e+01 -1.2143e+01
#>   3.3725e+00  3.7387e+00 -8.1309e-01 -2.1562e+00 -8.3301e+00  1.4185e+01
#>   6.9210e-01 -2.0255e+00 -1.2277e+01 -3.0098e+00  5.9826e+00 -5.7763e+00
#>  -5.1010e+00  5.3970e-01  5.9685e+00  4.8558e-01  4.0929e+00 -1.1917e+01
#>  -1.5037e+00 -9.6210e+00  9.6192e+00 -1.7651e+01  7.7463e-01 -7.9423e-01
#>  -9.6417e+00  4.6070e+00  1.1281e+01  6.3381e+00 -5.8177e-01  4.0138e-01
#>  -2.4994e+00 -2.2286e+00 -9.8558e+00  2.5175e+01  7.0611e+00  2.1639e+01
#>  -1.3800e+00 -5.1513e+00 -1.6259e+01  2.5677e+00  8.3732e+00  5.5927e+00
#>   8.4729e-01  8.0781e+00  5.8360e-01  1.0520e+01 -7.4458e+00  1.2502e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]