Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.7540  -2.0287   8.6957   6.2838   8.2548  -0.3801  -7.1020   2.8835
#>   -7.6431   9.0210  -7.1167   8.8111   0.0026  -7.8501  -3.0260   7.4157
#>    3.0124   2.3436  -0.9311   1.6793  -0.8888  -5.4673  -6.6721  -4.2742
#>    3.2138   4.5837  -6.0101  -6.4176 -15.0349  -3.8364   4.8951   3.2836
#>    0.9095  -3.7339  -6.5175  -2.2668  -0.2831  10.4166  -0.1837   0.7525
#>    1.0975   4.3797  -0.4136  -1.8459   2.5494 -16.1047   7.6727  -3.7225
#>   -0.4582   3.2298  -5.4885   2.4904 -18.7505  -5.2980   0.8231  -2.4121
#>    1.2275   1.1633   8.6643   0.0580   8.8530  12.4849  11.9277  16.5255
#>   -2.2935   2.9736   9.5863 -13.0030   0.7474   7.4428  -2.4226   2.1565
#>    0.5727  -5.2647  -5.1425  -3.0249  -8.5082  -3.5704  -5.5180  -4.7979
#>   -2.5740  -1.6674  -1.1578  -6.3719  -0.8676  25.2458  -7.3066   1.6180
#>    0.3256  -1.5935   5.6353   2.2402  -6.9674   0.8266   7.5809   5.9078
#>    0.5138  -7.4847   1.6443   7.9213   4.0959  -1.8777   5.9305  -3.6564
#>    0.1731  -2.7852  10.5299  -4.6337   4.6276   1.7118  10.2483 -10.4189
#>    1.7941  -3.5971   7.7159   1.3918  -2.1745   3.0205   6.5639   9.1390
#>    3.2649   6.6659   6.7818  -1.9596   0.2040   3.1017   2.9560 -11.7985
#>    0.5458   3.3660  -0.7184  -3.1946   1.9112  -5.7102  -5.5116  -2.2376
#>   -3.1135   1.5019   2.2571  -3.2395  11.0342   5.7858   5.9099 -11.4011
#>   -1.6205  -3.6083   3.6660  -5.4709 -10.0157  12.0901  15.3689   9.1229
#>   -0.9007   3.7062  -7.9542  14.6816  -7.6853  -4.8294  -1.0478  -6.0564
#>   -0.2187   0.5259   2.3065   4.4145  -2.3560  15.1009  -1.8690  10.1704
#>    1.1397  -2.0473  -3.6911  -2.5145  -0.4763   5.0938  -3.3959   4.4866
#>   -5.7127  -3.8662  -2.9239  -3.0447   1.8116   7.9448 -12.9671  -3.1741
#>    0.5248   2.7897   6.1374  -3.4621  -4.6873  -1.6263  -4.4175  -6.4763
#>   -3.7107   6.1587   2.7389  -3.0964  -0.5814  -6.6213   1.1715   8.0737
#>   -2.8685  -7.2257   1.0114  -3.4573  11.7799  17.9926   7.0164  11.8760
#>    6.2761  -2.2149   1.6719  -2.9661   0.1913  -1.4379 -16.9466   6.2541
#>   -3.3231  -0.7258  -0.7621  -3.7620   8.0627  -3.2855   0.4560  16.4697
#>    1.4732  -0.0071  -1.3449   1.2871   1.9485   6.5629 -10.0377  11.5145
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]