Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.6876  11.1725   0.1554   2.0304  -3.9165   8.1455 -11.1162  -4.7131
#>   -2.6330  -0.9252  10.2754 -12.2975 -11.7154  -6.6579  -4.4669  -5.9233
#>   -8.7403  -7.6876   6.8125   1.7440   6.8074  -2.8801   1.7181  -2.4560
#>    2.2691 -10.6863   4.3043   7.8457  -3.8763  -5.4312  -2.2657 -12.8237
#>   -3.8025  -4.1694   1.0350  -7.4590 -11.2604 -11.0793  -7.2513   0.4765
#>   -1.7286   6.1107  13.9471   1.8306  -4.5094  -2.9061 -10.4250  -7.4225
#>   -2.2933  -6.1192  -0.6131  -7.9606 -11.7232 -13.3265 -19.1155  15.4610
#>   -6.2316   2.8173   4.9702   5.1507  12.5420  -3.3688   6.4903   9.9981
#>    3.4529   8.3017  -2.8049 -10.0302  -8.0202  -7.2172  12.0648  12.1389
#>   -3.8308  -4.8808   1.7749  -3.0575 -22.9611  -9.5745  -6.8434 -18.2545
#>    1.2396   4.0122   8.3421   5.5378 -10.5029   2.8822  16.0368 -11.1556
#>   -0.4828   2.5704   6.7760 -16.6119 -12.3952  -9.6955  -5.8874  17.4948
#>   -0.7157   0.4521  -5.1253  -3.1298   0.4742   8.0160  10.6794   3.1956
#>   -2.3984  -0.9449   7.1017  26.3496  -7.7588 -17.7221  -5.3243  12.7218
#>    4.5374  -9.5317 -10.7282 -10.7350   1.2811   9.2573 -11.1621  11.3056
#>   -3.0846   4.5422   4.7454  12.9858   2.7612 -11.1878 -13.4864 -17.1334
#>   -4.4117  -1.0817  -3.8448  16.2725   3.7866  -6.0999  -4.9981  -3.0394
#>   -2.9870  -1.9876  -3.5642  21.1097   1.6073  15.4443  11.2988  -9.6763
#>    1.8323   9.6812   3.1524  -3.9025 -10.8820 -13.5238   7.7536  -5.1484
#>    1.2486  -5.0589   1.3278   5.1630 -16.6557  10.0539  18.6401   7.5549
#>    0.6948  -3.4221  15.7763   5.4637   6.7814  -9.3231  -7.2487  -2.1168
#>   -1.5092   2.6918  -3.5571  11.2345 -13.6476   0.6551  -4.4369  -4.2746
#>    1.2719   5.3730   2.0097   0.9589  -3.0840   0.1620  -9.2089   5.6280
#>    0.1834  -5.1447   7.7255  21.7990   0.2134   0.6886   1.2126 -12.7875
#>    1.5769  -1.3011  -7.7060 -10.5335  -1.9950 -10.6498   8.2585  -6.2783
#>    1.0674 -11.4560   3.4840   1.3086  14.6888   0.4511   2.5192   5.8171
#>   -3.7894  -6.1823  -2.3937   4.4715  10.1673   9.9306  -9.7735 -24.1733
#>   -4.6448  -2.2434   2.1362  -7.8307  -0.4055 -11.2642  -6.6439  -5.2447
#>   -3.5444   2.1350   5.3644  -1.5265 -13.2725   9.0263   9.8082 -10.6244
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]