Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -3.0021e-01  5.8721e+00 -6.0447e+00  7.0285e+00 -2.4308e+00 -4.2783e+00
#>   3.3098e+00  6.8433e+00 -1.7814e+01 -5.3236e+00  3.5537e+00  1.0543e+00
#>  -9.1442e+00  6.8067e-01  5.7746e+00  4.6333e+00 -9.8460e+00  6.2299e+00
#>   7.8186e+00  1.1838e+01 -2.4494e+00 -4.2859e+00  7.8282e+00  1.3165e+01
#>  -3.7069e+00  4.8983e+00 -4.5451e+00  3.8628e-01  8.1639e+00 -1.4086e+01
#>   2.2469e+00 -6.3389e+00 -2.8194e+00 -1.9343e+01 -3.9884e+00 -2.6301e+00
#>  -7.9327e-01 -4.3010e+00  3.4995e+00  4.9920e+00 -6.6805e+00 -4.0606e+00
#>  -1.0848e+00 -4.7949e+00 -4.3057e+00 -8.2884e+00  5.8540e+00  6.2237e+00
#>   2.5758e+00  2.2136e-01 -1.3643e+01 -1.1516e+01 -4.3648e+00  6.3958e+00
#>   7.8530e+00 -2.2834e+00  3.2728e+00  5.0045e+00  7.2312e+00  9.2654e-01
#>  -4.1286e-01  1.9806e+00 -7.5631e+00 -7.8988e+00  5.5658e-01 -5.3903e+00
#>  -4.3443e+00 -5.3124e+00  7.6460e-01 -1.2736e+00 -9.8731e+00  1.4567e+01
#>  -6.4775e+00 -7.2570e+00  9.1231e+00  1.5225e+01 -1.6030e+01 -2.9081e+00
#>   1.8282e+00  1.4137e+00  1.1582e+01 -4.5964e-01  6.5927e+00 -5.1264e+00
#>  -5.5490e-03  6.1303e+00 -1.7992e+00  2.6898e-01  6.4491e+00 -5.1281e-01
#>  -7.3623e+00  6.6079e+00  2.6151e+00  1.0427e+01  4.0051e+00  1.0038e+01
#>   2.6103e+00 -2.9701e+00 -2.8997e+00  1.1715e+00 -1.0011e+01  9.0070e+00
#>   6.0447e+00  1.0798e+00 -2.3577e+00 -4.0311e-02 -8.5928e+00  1.2732e+01
#>   6.0215e-01  9.7170e+00  7.3185e+00  1.2790e+01  3.0441e+00 -4.2901e+00
#>  -6.6052e-01  3.5000e+00 -1.1594e+01 -8.2050e+00 -5.5073e+00 -8.2514e+00
#>   2.6633e+00 -5.8143e+00  5.0388e+00  2.2269e-02 -2.2652e+00 -1.5147e+01
#>  -2.8730e-01 -1.7945e+00 -6.6258e-01 -9.9777e+00  1.7080e+01  3.9185e+00
#>  -1.2602e+00 -6.7238e+00 -4.4855e+00  8.3602e+00 -2.7328e+00  1.0917e+01
#>  -9.4046e+00 -2.9263e+00  9.1777e+00 -2.1245e+00 -8.8149e+00  1.1407e+00
#>  -4.1272e+00 -1.5306e+00 -2.3478e+00 -4.7164e+00 -1.1133e+01  1.0680e+01
#>  -1.7945e+00 -5.6864e+00 -3.1868e+00 -4.3079e+00  9.4924e-01  4.2483e+00
#>  -8.0562e+00  3.5870e-01 -6.0835e+00 -1.1852e+01  9.7088e-01 -1.3254e+01
#>   1.6999e-01  2.2220e-02 -2.5094e+00  5.8051e+00 -5.2723e-01 -6.0125e+00
#>   1.1919e+01  3.8833e+00  1.7256e+00  5.5737e+00 -1.0862e+01 -3.7101e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]