Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.9285  -1.1779   8.3169  -3.6717   2.1478  13.8795  -7.2530  -5.9560
#>    3.6127   5.7608   4.6235   1.8543   0.0465  -6.6146   1.4380   5.9157
#>    4.4367   9.3918  10.6323   2.2909   9.9661  -8.4901  -5.1961   0.1067
#>    7.3766   3.2555  -3.8127   5.1037  -0.1522   7.7729  -1.9690 -10.2461
#>   -2.9532  -2.6883  -6.7557  -0.1193 -15.2535 -11.5388  -1.7206  -5.4046
#>    1.1246  -7.2887  -2.1928   2.3167  -7.9368  -0.0835 -14.0233   0.8808
#>    4.4179   1.2844   2.0546  -9.0023   1.2979   4.2333  -7.2488  -4.4054
#>    0.1035  -3.7033  -6.9147  -4.4724  -3.5652   2.7812   6.0189  -2.1887
#>   -2.1341   3.3698   4.1177  -0.2790   5.9015   5.9827  -2.6241  -8.1410
#>   -0.6695  -2.4342  -0.0408   3.1362   9.9439  10.9345  -2.3098 -13.2161
#>   -3.4827  -6.7490   6.4056   8.7811   5.2250  13.1134  15.2013 -14.7152
#>   -1.1004   4.4736  -0.5746  10.5718   2.6061  -6.1674   7.1357  -4.2257
#>    2.6814   0.5488  10.9340  -1.6784   2.4316  -7.6311   1.2426 -16.7101
#>   -0.1841 -10.7876  -4.9357  -2.1033  -5.6753  -8.8643   4.7775  -8.7257
#>   -0.2619  -1.7303   8.9817   1.8475   2.4938  10.1083  16.3183  28.4320
#>    0.0039  -7.5365  -2.3921   2.2435  -8.2911   8.2069  -3.2459  -4.6744
#>    1.8853  -2.3171   6.8951   6.0099   4.0842  -5.4696  -7.0336  -7.9762
#>   -4.4807   3.0423   4.3289   4.8146  12.0885  11.9478  -3.8646   7.0378
#>   -4.7315  -2.9942   2.1663  -4.5487 -10.2249 -10.4668   1.0279 -21.0298
#>    2.5411  -2.3925  -8.4524  -2.1047  -6.2448  -6.8079   2.7928 -12.5212
#>   -4.7642  -9.7784   4.5228   7.0804  -7.7658  17.4524  -2.4354 -22.5578
#>   -2.2559   5.8539   6.3473   5.9928  -7.6165   1.5899 -11.9487   8.0338
#>    2.0673   0.2127   1.4399   4.0521  13.9240  12.0592 -10.7811  -1.4264
#>    2.1423  -0.2435   1.2187  -4.4068  -1.6246   2.4896   0.6881  -3.3649
#>   -0.9945   5.6398   1.6974   0.4095   6.2746 -14.7112   3.1838  20.6946
#>   -0.7440  -9.9958  -0.9693   6.8122  18.3713  -9.5764   8.2665  -2.2972
#>   -2.3256 -12.1935  -3.9373 -11.4969  -1.3414   2.6484   9.6371   0.4271
#>    1.3091   1.1460   8.3381  -5.0773   9.6885  -3.3753  13.4502  12.9395
#>    3.2995  -4.9078  -3.1535  -4.0490  11.5663   0.7195  14.4726   7.3073
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]