Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.1165  -2.8477 -13.3407   6.6262 -16.1702  11.4118   1.8718  -3.7992
#>    4.0602  -9.9178   5.6219  -6.7661   6.0325  -8.9425 -11.9045   4.0973
#>    2.3423   7.2430  -0.8189 -16.4453  -2.0892   1.5339   7.2613   9.8535
#>   -1.7118  -6.5207  -6.3669  -3.8880   2.4247   1.4277  -7.7479  -0.6743
#>    5.0506   6.9772  -3.1743  -7.6268  -2.4812   4.0942   3.3271  -9.4298
#>    4.0314   1.1560   0.3048 -13.8698  -7.4882   7.5332   2.4947   5.9849
#>   -3.8269   7.0883  -7.7780  -1.2970  -0.1600  -5.9192   5.9686  10.6764
#>   -2.7813  15.6705  -2.8946  -0.0525   6.2887  -0.8519   6.4635   1.0085
#>    2.4836 -16.3761   2.0367  -3.2372  11.5200   0.6261 -12.2909   6.7558
#>    2.7320  -3.8128   0.0975   3.1344  16.5966   2.0757  -3.9115 -13.0485
#>   -2.6684   3.7864   0.2256 -21.3801 -20.0875   5.6046  -6.6191  -6.7016
#>   -0.4237  -1.7721 -15.5630  -8.4852  11.4843   4.3074 -12.3618  -1.9022
#>   -0.4637  -3.0607  -5.5096  12.5634  17.9885   7.5310   5.2088   0.4172
#>    3.7060 -10.7560   0.8355  -6.2445 -16.8405  17.1575 -21.4630  12.5420
#>    3.9414   4.6583  -9.4487 -13.0932  -9.9300  -9.8128  -1.2166  -3.0561
#>   -1.8606  -1.5488   2.8841  13.1180   3.1542   1.7916  -9.8982  -0.7847
#>   -5.4435   6.2488   1.2354   9.3723 -17.7432 -10.4864   1.0214  -0.8167
#>    5.2662   0.9786   3.3447 -15.2304   6.8574   0.4944   1.7030  -1.5854
#>    1.4303  -2.7918   7.7457   2.6509  13.2384  -3.7680  -6.2004  -3.2511
#>    5.1772  -6.5204   5.3930 -11.9590  -4.6313   5.5563 -11.7547  -8.2306
#>   -2.5646  -8.0806  -1.9302   8.4564  10.0202   2.5066  -6.6131  -3.0547
#>   -8.5844   4.2541   5.1356  13.5727  -0.8156  14.0573  -4.6539   4.6157
#>   -1.1303   8.3012  -1.4780  -2.5475 -19.5294 -19.6195   0.9867  19.3337
#>    6.9771   4.1836   6.7620  -4.2769 -14.1925  -0.0377 -19.1008  -1.0721
#>   -2.0421  -7.1436 -10.8212  -1.5154  14.5999   4.0495  -9.4473   0.1558
#>    3.2732  -6.8922   9.1558   0.7849  13.0817   6.0651   4.0783  -2.3573
#>    3.3621   8.3177 -15.9265  -3.8579 -13.9566  12.4740   1.7363  13.9325
#>    3.7248  -3.3220  17.3003   8.5946   5.5563  -3.7909  -0.9501   4.7140
#>    4.5168 -10.8850   9.2252  -2.7694   0.0647   4.4564   3.8685   9.0632
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]