Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -3.8708e+00  7.2834e+00 -3.2233e+00 -1.9707e+01  6.5388e+00  8.6249e+00
#>   1.6598e+00 -1.1909e+01  4.5622e+00  4.0448e+00  9.2991e+00 -1.7650e+01
#>  -3.0418e+00  4.3691e+00  7.7855e+00  5.0048e+00 -1.2200e+01  1.3831e+01
#>  -4.3818e+00  3.0175e+00  4.7117e+00  1.3477e+01 -1.6623e+01  1.0589e+01
#>   1.8973e+00 -1.1616e+01 -3.0260e+00  3.1438e+00  1.7296e+01  7.0253e+00
#>   5.0913e+00  4.5414e+00 -1.2032e+01 -2.0193e+01  1.6070e+00 -8.1409e+00
#>  -3.0371e+00  6.4498e+00  1.1763e+01 -7.5787e+00  1.1761e+01 -1.0366e+01
#>   3.3024e+00  7.4227e-01  5.8479e+00 -3.9654e+00 -6.0132e+00 -7.9845e+00
#>  -4.8051e-01  4.4636e-02 -1.2558e+01 -1.1747e+01  7.5306e+00  2.1437e+00
#>  -7.5687e-01  7.8661e-01 -3.5931e+00 -6.9322e+00 -1.0033e+01  2.3058e+00
#>   4.9762e+00  7.2850e-01 -7.6972e-02  1.8078e+01  4.8836e+00  4.6634e+00
#>  -5.8651e+00  4.5028e-01 -7.4089e+00  2.0112e+00 -1.1844e+01 -3.7957e+00
#>   1.3284e+00  8.9805e+00  7.7668e+00  1.2388e+01 -1.5561e+01 -7.1615e+00
#>   8.3428e+00 -9.9974e+00 -3.6877e+00  4.7225e+00  1.5179e+01 -4.7010e+00
#>  -3.2821e+00  1.4455e+01  8.3354e+00  2.4332e+00  4.3178e+00  1.3002e+01
#>  -3.0890e+00 -9.9786e+00  5.9168e+00 -3.6933e+00  9.4576e+00 -5.2659e-02
#>  -1.3251e+00 -1.9236e+00  1.4812e+01 -5.2738e+00 -4.3582e-02  4.8969e+00
#>   4.7952e+00  4.6176e-01  8.0178e+00 -3.5602e+00  1.7836e+01 -5.4175e+00
#>   1.7638e+00  7.1745e+00 -2.1136e+00 -1.3493e+01  1.0145e+01 -1.5558e+01
#>   1.9990e+00  1.1958e+01  8.1478e-01 -1.2304e+01  9.7623e-01 -8.9021e+00
#>  -3.1084e+00 -7.3787e-01 -2.4366e-01  4.1672e-01  3.3730e+00 -4.2513e-01
#>  -5.9233e+00 -7.2364e+00 -3.3975e+00  1.0635e+01 -2.5491e+00  2.1382e+00
#>   1.0575e+00 -3.3765e+00 -5.4334e+00 -4.0380e+00 -6.6195e+00  1.7710e+01
#>  -5.6482e-01 -1.3966e+00  1.7276e+01  4.9108e+00  1.4125e+01  5.7504e+00
#>   1.7115e+00 -3.5550e+00  5.3521e+00  1.0383e+01 -6.5531e+00  4.1711e+00
#>  -3.5526e-01 -8.8658e+00 -7.8314e+00 -3.3360e+00  1.2422e+01 -4.1071e+00
#>   8.0140e+00 -3.1349e-01  6.9467e-01  5.4895e+00 -4.3903e-01  9.9884e+00
#>  -3.8857e+00  5.5199e+00  8.4898e+00 -7.6777e+00  9.5171e+00 -1.3684e+00
#>  -2.5096e+00 -5.4124e+00  1.1076e+01  5.4601e+00  3.2136e+00  5.0008e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]