Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.3648   0.2726   1.2061   0.0517   7.6861  -1.6124  14.9034   6.9403
#>   -1.9277   5.2003   5.5396  -3.5241 -21.4855 -13.2491  -4.1647 -12.3212
#>    0.9112  13.8089  10.5016   5.2561   3.1829   8.0805  13.4542  -0.5625
#>   -1.0355   2.2722   5.9338  -5.9774 -17.3005 -12.6274  -6.3361  -3.2522
#>    5.5522   1.7536  -2.6705   9.9698  -0.2024   0.0433   2.7357   4.0870
#>   -6.9706  -2.9027  -6.1930  21.1012   0.2378 -18.1547  18.5481   9.6852
#>   -4.3575  11.2010   5.4481   5.8959   2.3125   6.9468 -13.8365  12.8373
#>    3.5355  -0.8737 -10.0031   1.2952  -5.1471  -3.3978  -9.1442  -1.7570
#>    8.1661   4.9773  -7.1823   1.9240 -16.9478 -17.8782  -6.2861  -0.9422
#>   -2.6030 -22.7903 -11.2757   6.2753  -9.6099 -14.8295   4.7343   6.9982
#>   -5.3549   2.1657  -3.3021  -1.9426   4.4495   6.9791   5.6756 -13.8728
#>  -10.5231 -12.1437 -11.1408  -2.4576   3.3822   4.0000   0.6003   3.2193
#>    1.7135 -11.7676  -5.6406   2.9753   0.9473   5.2835  -4.0383   4.4307
#>  -10.8540  -2.0544  -9.1392  -6.1805   0.9451  -3.4652  -3.7963   5.7923
#>   -3.0884   2.8077   0.2664   9.4084   6.4757   4.2038   3.7872   1.7322
#>    6.6386   3.8178  12.5466  -9.4572   5.7934  -2.8585  -1.8756 -10.9501
#>    3.4519 -12.2869   2.8215   5.7122  -5.0532   6.2679  15.9006  -8.0791
#>   -3.5719  -4.9474  -9.7594   2.4000   2.5781   4.0856   6.6029  22.7592
#>   -2.5657   1.7292   2.0717  -0.5805   5.3799   7.1031  -8.1447   0.5414
#>   -0.9623  -1.1542  -3.5651  -0.9754   4.9077  -1.1221  20.5906  -7.2158
#>    2.7666 -13.5720   1.1433  -3.4127   2.6166  -6.8831  -1.3434 -16.3249
#>    0.8187  -0.6225  -2.1505  -5.8914   7.7105  -2.5421   2.2309   1.0939
#>   -4.1455   8.1596  -0.4277 -18.5123  -9.1524   7.1006  -3.8420  -4.5165
#>   -1.4933  -8.3973   7.7977  -6.8046  11.7283  13.4870 -11.9564   7.3608
#>    1.7180  -5.1782 -18.2052 -23.7729  -7.6521 -14.5893 -10.7001  -7.7381
#>    4.9121   5.0734  -7.9034 -20.2446   7.4130   7.4844  -3.5911  10.4452
#>    3.2840   7.9470   5.7675  13.6539  15.1159   5.6723  -6.3620  10.2577
#>    7.9372   4.8620   4.8294  -1.4041  10.6401   2.0151  18.2784   1.8501
#>    3.2313   5.6353   7.1634  15.1469   9.8823  11.6760   1.2554  -9.1242
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]