Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.0858  -9.6636  11.4350   9.2217  -7.0134   6.2987 -11.6979  -0.2638
#>    1.8460  -4.1198  -4.2168   0.3048  12.3601  12.2236  -2.4386  -4.2768
#>   -2.8778   6.2990 -10.8131  -3.1342  -6.0364   1.3518   8.1377  -4.2262
#>   -2.4273  -4.6229  -0.7334   2.1519 -10.8926  -9.9349  -3.8018  -2.6220
#>    0.4323  -2.4802   1.3777  -0.4900   5.0598   5.8836   7.7597  -7.1967
#>   -2.4460  -1.4758  -5.6540  -3.3577   3.7683 -21.3449  14.3630   5.4377
#>   -0.7553  -4.7226   6.9819  11.9229   8.5108  -6.7903 -13.9108  10.8797
#>   -5.2929   4.0665  -6.9452  -5.3651  -9.2635   1.4342   2.5658 -10.6917
#>   -3.2824  -1.3483   6.2549   3.9953   6.6793  -8.5009  -2.0621   1.0144
#>    9.5102   5.0186   3.5172  21.2435   9.0893  -9.6084   1.4718  -6.7050
#>    1.0122   4.0781 -15.3022   5.5854  11.6209 -10.2009  -0.3944  -4.4954
#>   -0.9191  -1.2354   5.9514  -0.4916  -8.9461   7.9249   0.7508   3.7082
#>   -1.7593   2.9420  -2.9935  -0.0169  -1.9499  -1.1221   3.4061 -12.9560
#>   -0.5874  -3.6706   2.4562   1.9513   2.5657  -4.3842  -1.2437  -5.4175
#>   -0.4115  10.2639  -0.6545  -8.5860  13.9119   2.2332  -2.7978 -10.3695
#>    3.6592   4.0977  -5.0712  -9.4429   8.0027   3.5502   4.2005   4.5506
#>    0.1291  -3.9940   8.0124   6.5953  -8.8546  -3.1643   6.6364  -6.3730
#>    8.2292  -1.2471   0.0607  -3.1471   6.4481  11.3855  -0.3026  -3.1466
#>   -3.3923  -3.4766   1.3566  -1.0758  11.0122   1.9885  16.0912   1.9082
#>   -2.6802  -5.0612   0.4790  -2.3463  -3.9849   0.4050  13.0325  -3.8474
#>    2.8640  -5.1031   1.9571  12.1453   0.3799   3.5121 -15.5428  -3.0356
#>   -1.4389   1.2039  -7.8738  -8.6576 -14.3866   5.1935  -5.7285  10.1294
#>    1.3064   8.6011  -2.4756  12.4218  10.6884  -6.2319   6.8616  -3.9808
#>   -0.5561  -1.5026   0.7435   1.8033  -7.7918  -2.9370   7.4343   2.8405
#>   -3.4855  10.7410  -0.0060  -0.2520  13.6924  -1.6807  10.6933   7.4928
#>   -0.5063   1.7782  -0.1338   6.0271  14.4088  -2.3855   0.3134  -3.0626
#>   -3.8682  -2.7251   1.5601   0.3109  -5.7191  19.2731   3.2442  -6.5142
#>   -3.7777   3.6697 -10.5550   1.6254   7.8846  -5.3321  -1.3285  10.7415
#>    3.0807  -5.0706  12.9638  18.6107   5.3092  -0.8497  10.3703   4.8753
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]