Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.3176   1.3201   9.8588   7.3900   3.9447  -1.4838  -1.9540   4.4520
#>   -0.2129   2.4826  -4.3958   3.9829  10.8466   0.0929  -3.0165   1.0312
#>    2.3499  -4.0355   1.6481  -3.7956  -7.6514   1.4622   5.4711  -5.7765
#>   -0.9659  -2.2856  -0.1647  -8.1096  19.8790 -12.4209  -5.8986  32.5410
#>   -5.3917  -0.6551  -6.1652   0.0525  -0.2648   8.0516   2.4846  -3.1541
#>   -2.7679 -11.3354  -0.3493   7.0753   9.2692   9.9384  10.5780  -5.9762
#>   -2.2658   1.8418   2.3498  -8.8808  -1.6856  -9.9888 -13.6581   1.2025
#>   -0.6892   7.7398   1.9558  -1.2972 -11.4883  -8.2764   6.8219 -15.5495
#>   -0.8102  -5.2269  -4.1301  10.8462 -15.1095  -3.9837  -4.8187  -3.9280
#>    0.2644   5.4710  -1.4384  -5.7100  12.4629  -0.6012  -6.9679 -17.1010
#>    0.7047  -1.3854   0.0838   7.3984  -7.7985 -15.5590  10.4427  -0.2801
#>   -5.8316  -1.1891   2.9586  -0.2096 -18.4960   1.7600  14.8140  -0.9865
#>    2.7834  -3.7020   1.0773   9.9892   8.8311  -4.5609   4.9670  -2.9019
#>    3.0562   0.1503  -4.7390   6.1384  13.0412   3.5860  -0.7172  -1.3342
#>    1.8917   3.0060   3.6497   2.0561  -7.6453  -5.3196  16.7624  -7.3403
#>    0.8107   7.8791  12.6621  10.2943   5.0880   2.4949  -3.8305  -8.4788
#>   -0.2827   0.3998  -0.7469   3.4472  -4.1728  -9.7450   5.4255 -28.7336
#>    0.7879   4.9258   6.0824  14.3525 -13.5860 -14.6562 -12.6422   7.4578
#>   -9.0012  -3.3001   6.5808   5.9549   3.1156 -17.8496   2.6209  -2.9906
#>    1.4610  -6.3957  -1.6605   3.6956   5.8067  11.7449  21.2662  -7.2997
#>    0.2522  -2.3385  14.0292  10.8508   2.3930  -1.1978  -3.7370  -1.9547
#>    1.8432   2.8835  -2.7499  -4.0536   0.6415  20.1322  -8.6372 -14.1020
#>    0.5157   3.1465   1.1577   4.6904  -0.7526  -6.4070   6.1286   7.4642
#>    1.0027  -1.2267   4.6537   5.7695  10.0737  -0.3578  -9.0975   1.5513
#>   -0.1115   3.7557   3.7496  -1.4795   9.0758   5.0102   2.0557  -6.9581
#>    2.2449   4.2382 -12.8616 -14.6523  13.3552  -6.8264 -10.0834  10.4895
#>   -6.2687  -4.8958 -14.1157 -15.1154   4.2326   4.3705  -6.5907   0.1030
#>   -1.6795   3.4018   7.2937  15.0076  -7.0408 -19.1144  -5.1341 -10.3773
#>    0.0938  -2.8423  -9.2902  -5.5731   6.2968   2.1833  11.9364  22.8522
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]