Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.0871  -9.6214   7.5141 -11.2933  -2.4140   5.2864  -8.8742   4.4597
#>    4.1577   2.7253   5.9360  -3.4242 -10.2469  -1.0881 -19.8280  -3.4554
#>    1.7022   0.1483   3.0436  10.4502 -24.4698   1.0826   7.3361   3.5896
#>   -1.2625   3.9724  -1.3095  16.0913  -2.2085  -4.2060  -9.7798  15.8315
#>    0.5921   7.9903  11.7059  -0.2897  12.4142  -6.1883  -5.2586   2.0773
#>    1.5151  -5.7816 -12.8260 -20.3340  -1.6557   7.6308  -4.6455  -4.4529
#>    0.4830   5.7448  -9.2286  -7.8512   0.6630   3.3512 -10.1900  15.5690
#>   -3.8960  -2.3562   5.1621   2.2345   8.5688   5.3868   9.6959   6.6933
#>    0.1714   2.3227  -6.0272   0.9088  -0.2334  10.8073 -10.1961  -4.7837
#>    0.5137  -0.7788   8.3457  13.7270  -7.8618   1.0458   1.1277   4.3513
#>   -3.4936   9.0183   2.1394 -26.1658  13.0644  -0.6963  -6.6024   4.1550
#>   -3.3275  11.3062  -2.2643  -5.2570   2.7082  -1.5457  -4.3180  -8.6792
#>   -1.3430  -4.6155  -4.3233  10.2217   6.4908   5.7214  -2.0447   0.6347
#>   -2.8049   9.4536   4.2416   7.8897  -2.4201  -3.1836  -1.3627 -20.2497
#>    2.5994  -6.8164  -0.9030   2.5320 -11.4654 -19.3501  -6.6198  -8.3474
#>    1.8168  -8.8308   0.1486  -3.2483  -0.5234 -12.1064  -7.3653  -4.8519
#>    5.2501  -8.9352   5.1057   1.7480  15.7883   5.5153 -15.3762 -10.1677
#>    1.1032  -0.8444  -0.3247   2.2246   1.5947 -10.1730  -6.0686  -8.3259
#>   -3.2739   2.6148  -9.8895  -7.3635  16.7047 -11.7062  -3.2932  -0.4064
#>   -3.0305   5.2358   4.1630 -22.0216   2.8120  -5.1538   2.0108   0.5335
#>   -2.0038  -5.7869  -5.5812   5.3190   0.5428  -1.3568  17.6269   1.7837
#>   -1.3283  -6.7990  -1.7662  -3.9771  21.2994  13.4070   3.7509  18.4754
#>   -5.4956   8.2801  17.7489  -4.4433   9.3634  -6.9939  -6.9148   6.8715
#>    5.4190  -2.8496 -13.9606  -8.3333 -10.1180  -2.5602 -30.9835 -12.3645
#>    4.7447   2.0280  -0.7571  -4.2245  13.7238   8.4820 -23.1614  11.1453
#>   -1.5172  -4.5484   1.5666  -7.4676   7.2873 -16.6771   4.7926  -6.4781
#>    0.6461  10.4118   6.2946  -3.4522 -10.6261  16.4878  -2.7827   2.7925
#>   -2.8098   4.6010  -9.2990  -3.8307   1.5748  -0.8862   2.2897  -3.2155
#>   -0.5845  -2.5847  -9.3477   2.0472  14.4177  -0.0527  -1.9929 -10.1198
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]