Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.1630  -8.0486  -6.5548   5.1953  -4.8256  -2.1856  -7.0777  16.8079
#>  -10.9443 -10.9732  -2.0834   6.8456 -20.0724  -5.9316  -7.9971   2.4694
#>   -3.3056   3.4286   3.2570 -13.4686  13.8622 -14.1344 -10.1280  10.2739
#>    1.5237  -1.6879   1.8901  -0.2658 -16.3421  23.1631  -4.1483   4.3026
#>    0.5983   0.1775   4.4258   6.2256   0.2397 -12.0822  -3.7352  -1.2158
#>    0.9094  11.7908   0.1904  -5.8655   7.3595  26.4955  -8.3877  19.0604
#>   -5.9033   4.3775   2.5837 -18.9506  21.2566 -11.8606   2.3659   3.2445
#>   -2.5780   2.8150  -0.2636   0.7090  -8.5389   5.1329 -24.0503  -1.1333
#>    5.8892   5.4291   8.1779  11.0600 -14.2364  -6.1276  -7.0295  -6.9599
#>    2.5624 -11.7537  -2.0200  -7.0960 -13.2017  -8.4189  -3.1143  -2.9559
#>    1.5904  -2.8142   3.6180 -10.4801   7.6511  -0.0987  -4.7305  -8.6730
#>    1.6645   4.4164  -4.8236   6.3036  13.5825  16.9047  19.3458 -21.8273
#>    2.5141  -4.2266  -4.9087   5.0033  12.6967 -15.8811  -4.4165   3.4306
#>    5.3892   1.4573   1.2081 -11.0650  -4.8079 -15.5219  -5.8802 -14.8694
#>   -0.3207   3.0748 -14.3956  -2.8450   9.6549  -0.5014   8.9391  12.2438
#>    3.4524   1.3462  -4.3199 -18.9404 -12.5827  -5.0218  -2.0432   1.2342
#>    3.5886  -6.2720  -3.8689  -7.5920  -0.8027 -14.0657   6.4842  -9.2667
#>   -7.3858  -0.9269  -5.1636  -1.3072  11.4683 -17.1784  -5.2885   8.7553
#>   -7.9036  -4.6332  -6.0611  -5.2997  26.3667  -3.5180   6.0801   8.2331
#>    2.9227   3.3609  -7.5656   2.7626  -2.8807   6.2148  10.5387  -3.9580
#>    2.9565   0.2601   1.7218   3.8457  -7.7939 -10.7251  -6.2765  -5.7697
#>   -4.7795   3.3680 -10.6519  -2.9417  -4.6275  -4.6973  12.2969  -6.6184
#>   -3.7715   3.1981   7.3901  -8.8852  15.0703   2.6458  -0.9383  -0.3736
#>    8.1975  -3.6626  -7.6658 -17.9105   7.9487  -4.1133  -4.6777  -0.2380
#>    1.3461  -1.5420  -0.3142  -1.1313  10.7415  -4.8220   8.1401 -12.0396
#>    0.4066  -3.4382   7.9853  12.0495  -7.7629  -6.0211 -10.1226  -8.4840
#>    1.9329   1.5109  -0.3555 -16.2343   9.0411  -9.9030  12.6333   7.1767
#>    2.4954  11.0818  -3.2367  -0.6347   0.2523  -6.7341   2.7961   5.3513
#>    0.2672  -6.4669   2.3486  -9.2753   8.7072   6.0221  -7.7930 -11.6151
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]