Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.2985  -0.3849  -4.3161 -11.4835   4.9223  -3.0197   6.0074 -10.8100
#>   -1.3989   6.6490  19.4224 -20.6153 -16.9465  -0.4997   5.3068   2.2700
#>    2.7098  -8.0499   3.6061  -6.5519  -6.5360   9.3994   0.4541   5.0253
#>    4.7883  10.8986   4.6260  12.3264   1.9925  -5.0113  -1.3592  12.9226
#>    2.8563   9.4538  -2.8986  -7.4218   0.0719   3.7765   6.7093   3.9520
#>   -2.0966  -6.7597  -1.3538  -0.2518  -1.7601  -7.3322   2.1218   0.2031
#>    2.3037  -6.6550   9.5884  -2.1647  -2.7438   7.1228 -10.9109  -4.3936
#>    1.1223   8.5485  -8.9222   5.1361  -8.9736  -5.6134  -3.7409 -10.8711
#>    4.4523  -0.7963 -10.7397  10.0764   7.4142   1.2262   9.3282   5.6553
#>   -5.6222  -2.6729  -3.9194   5.5527  -1.0943   8.3640   0.3725  -6.4498
#>    4.6771  -0.0935  -7.3626  17.8088  -8.9726   5.4179   1.0368  -1.3694
#>    4.8822   3.1923   1.6647  -0.1121  -0.8760   2.9965   9.6650 -12.3654
#>    6.5873  -3.3553   9.1851  -5.3820   3.8210  11.0783  -4.1371   3.3425
#>   -7.8319   9.1221   9.7875  -4.8771  12.3396   5.6363  -5.4855   0.0096
#>   -3.7975  -0.7395   7.1096  11.1247   1.5506   0.3524   8.6101 -12.1496
#>   -4.3612   2.1537 -12.3728 -16.1321   2.3904  12.9931   1.9644  -8.8978
#>    5.4309   1.9537   6.2754  22.6106  -3.5657  -1.1614   2.6224  15.6163
#>   -1.1764  -1.0756  13.9348   2.8177  -3.0229  16.4079   5.9312   6.5458
#>    4.0941 -15.3828  -4.1399   7.4812  -3.4754  -0.7546  -9.1149   5.1887
#>    2.5032  -6.3908   4.1713   9.0716   6.6498   2.3796   4.0920 -17.0411
#>   -0.4367   5.4299   4.0881   1.1233  -7.7070   6.3710   3.7487  -2.3782
#>   -3.7509  -0.8865   1.9487   8.0436   9.3632  -8.0957 -14.8535   0.1559
#>   -6.5305  14.1022   2.8687  -9.7593   4.0837  -3.8711  -0.5146 -12.3155
#>    4.5918 -10.0348  -0.5748   1.0764  -6.0501  14.6936  -1.3043  -5.3261
#>    6.0264   5.5274  -4.9226   0.6728  -5.1479   0.3195 -11.1743   3.8141
#>   12.9295   1.0259  -3.2832  14.0493  -2.8471  -5.4675   2.3978   6.8462
#>    1.1036   3.1332   5.4422   0.6261 -10.6511  -0.2199  -3.8685 -13.8097
#>    2.8222  -5.7891   4.6352   8.8316  -8.0424   0.9178   7.2226   2.5181
#>    2.0101  -1.5768   2.2225  -8.7351 -11.6278   5.5509  -2.2778  -1.3952
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]