Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -1.1042e+00 -1.9546e+00 -5.3950e+00  2.9108e+00 -1.8851e+01  3.4669e+00
#>  -2.7612e+00 -3.1065e+00 -5.2817e-01  8.6737e+00 -1.3185e+00  1.0678e+01
#>   1.8606e+00 -4.1960e+00  7.8281e-01 -1.2613e+00 -3.6992e+00 -1.4807e+00
#>  -4.3771e-01  1.0113e+01  2.9197e+00  1.1209e+01 -4.0649e+00 -9.7968e+00
#>   5.4428e+00 -7.2509e-02 -9.2737e+00  8.1763e-01 -8.9594e+00  6.5423e+00
#>   1.9228e+00 -3.7010e+00 -1.0343e+01 -8.9250e+00  2.3915e+00 -2.0580e+00
#>   4.3744e+00 -7.0941e+00 -1.0287e+00  6.2726e+00 -2.5144e+00 -3.9828e+00
#>  -1.7411e+00  5.9368e-01 -1.3536e+00 -1.4749e+01 -1.0056e+01  8.4392e+00
#>  -3.3410e+00 -2.3610e+00 -3.9337e-01  1.0337e+01 -9.4005e+00 -9.8820e+00
#>  -1.6571e+00 -2.0122e+00 -5.6134e+00 -3.2520e+00 -4.8382e+00  1.7318e+01
#>   1.1902e-01 -6.7966e+00  2.3077e+00 -3.7614e+00  1.0359e+01  2.8023e+00
#>  -7.8228e-02 -4.0182e+00  2.0402e+00  7.9966e+00  6.4588e+00 -7.4187e+00
#>  -7.7974e+00  1.0951e+00  6.5030e+00 -7.2399e+00  9.7117e-02 -1.3519e+01
#>   1.9601e+00  6.1735e+00 -5.6774e+00  6.3175e+00 -1.3486e+00  5.2862e-01
#>   5.1948e+00 -3.3814e+00 -8.6801e+00  7.8275e+00  1.5910e+01 -5.2394e+00
#>   2.2054e+00  4.6727e+00 -4.3403e+00  4.6425e+00 -3.6982e+00 -6.6871e+00
#>  -1.2848e+00  8.0486e+00  1.2439e+00  8.7120e+00 -1.2125e+01  6.6657e-01
#>   5.0085e+00 -3.5722e+00  5.0697e-02 -6.3925e+00  1.4302e+01 -1.5100e+01
#>   4.8034e-01 -2.7189e+00  1.4303e+00  2.3973e+00  8.7728e+00  1.5442e+00
#>  -2.6985e+00 -1.4106e+00 -3.8911e+00 -6.7585e+00  1.2570e+00 -6.8556e+00
#>   4.1218e+00 -1.0952e+01  1.0797e+01 -1.1055e+01  6.8874e+00  1.2903e+01
#>   2.2406e+00  4.5748e+00  1.7158e-01  4.4351e+00 -1.4304e+00 -4.7047e+00
#>   1.3871e+00  1.4201e+00 -1.1096e+00 -2.5759e+00 -1.4925e+01 -1.6058e+01
#>  -2.8244e+00 -8.2303e+00  4.5263e+00 -2.6240e+00  6.9057e+00 -3.9447e+00
#>   2.6407e+00 -5.2865e+00  1.4398e+01 -8.7224e-02 -9.5304e+00 -1.0700e+01
#>   8.0203e-01  1.1271e+00 -1.6899e+00  6.9571e+00  4.4216e+00 -1.0139e+01
#>   2.0867e+00  6.3466e+00 -7.4202e+00  1.1764e+00 -1.5686e+00  3.0984e+00
#>  -2.8240e+00 -1.4145e+01  4.3970e+00  5.5147e+00 -3.7856e+00  8.8229e+00
#>  -9.5669e-01 -1.8200e+00 -2.0185e+00 -3.9206e+00 -9.3718e+00  1.7988e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]