Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.1567  -3.9723   3.7059   0.7331 -24.3742 -13.4914   0.7067   3.0803
#>   -0.8226   0.9444  -0.2651   2.3424  17.2151  -3.4419 -11.6949  -6.0035
#>   -0.3246  -3.5457  -2.3067  -2.1624   8.9477  12.0073  12.2133  11.5436
#>   -5.1362   2.6682  12.7613  10.5053  -5.1564 -10.0201  25.1063   4.8382
#>   -2.8158   1.9029   5.6077  -3.0277 -13.4372 -14.5312  -8.0440   0.7948
#>   -0.7957  -3.7425   2.0250  15.3515   3.0393  17.5745  -1.5759  15.0576
#>    3.7811   2.7299 -11.5947  -9.0259 -16.4135   2.8940  14.4225  -7.3134
#>    4.2412   0.2474   3.3924  14.4221  -0.0141  -3.6562  -3.9885  14.3063
#>   -3.5013 -10.6809  12.6981  -3.6788 -18.9103 -12.2159  -0.8605  16.7706
#>   -4.0801  -4.3800  -7.3984  -3.8849 -10.4679  -2.8112 -18.9416  -1.2434
#>    1.5350  -3.2722   7.7618   6.5865 -13.3808  -4.2435   1.3991  -6.3784
#>   -3.6069   2.5787   8.0726  -1.8627  -9.3236  -4.5791  -5.3408  -2.2954
#>    0.9791   3.6514   4.2279  -6.3617 -25.3889  -2.8945  -0.4378   6.7468
#>   -5.2433   6.9180  12.8160   4.4217 -14.4520   3.9650  17.0209   3.0855
#>  -10.0043  -5.3185  13.6162  -9.6765 -14.7978  -4.3664   3.5472   2.4228
#>   -2.1593  -5.2039  -2.0140  -3.9231  -6.8254   7.9980  -4.7496  -2.0406
#>    2.2655   2.8577  -8.8574  -7.1860   3.1096   0.3358 -13.9730   4.9946
#>    1.0102  -8.1073  -4.0804   8.0153  -5.2338  -3.5811 -11.4459   9.7403
#>   -5.3278  -5.0952   5.9620  -7.6742   2.7926  14.1808  19.1438  -3.1492
#>    1.7764   4.3170   2.9282   2.4738   8.2593  -5.4065   7.1658  15.2858
#>    0.7616  -7.6407  -5.3650   4.3803  13.1430  11.1419 -19.8366 -14.3603
#>   -0.6361   1.1763  -5.6087   2.5333  -1.0400  -7.0944 -13.2114 -14.2406
#>   -2.9346   2.4507   2.7625  -0.5688  -2.5512  11.5336  10.4606  17.1867
#>   -3.9939   3.5280  -2.4411   0.1821   0.0337  -9.7259   9.3316 -11.5841
#>    3.9903   4.0175   7.0886  -9.9730   1.7717  -8.8317   2.7375 -14.6304
#>   -7.2633  -1.1694   6.7178  -7.4150 -15.0896   1.6116   4.7827  -6.5290
#>   -3.3314 -14.4717  -1.8698  10.2139  -2.0324 -17.4194  -6.7235 -10.4298
#>   -2.7123  -0.8017  11.6229  13.3463   2.2265   1.0001  12.2750  -0.6462
#>   -0.3836   0.7767  15.2005  -3.1557   4.9260  -2.1765  22.2679   0.5487
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]