Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.5176   2.2150  -0.8525   6.1151  -4.0872   9.3212  -2.0020   5.6630
#>   -0.2863  -3.0740  -2.5491  -0.4353  -3.5870  17.2017  11.5740   4.6048
#>   -0.3253  -1.2595  -2.8852  12.2931  -3.3419  -1.7019   3.1748  10.2298
#>   -1.4872   1.5606  -3.3040  -3.6211  -0.0893   6.7127   3.2093 -15.4303
#>    1.5571   2.4811  -0.9439   9.5855  10.2512 -10.7242   2.4509 -11.7410
#>    0.7513   4.7210  -4.3265   2.6141   1.0906 -20.6340  -4.4524   4.6838
#>   -1.2731   5.0224  -7.7370   0.2228  -8.6613   1.2110 -24.3799  -0.4753
#>   -2.0117  -1.6429 -11.1670   3.4132   8.3252  -0.0406   8.7273  -7.1106
#>    0.4511  20.6644   0.0467  18.5586  18.2322  -7.0490  -8.2402  -6.1660
#>    0.2857  -7.8691   5.3240  -3.5914  -1.5972  -5.1441  16.9856 -15.6431
#>   -0.8082   6.2956  -2.3364  -0.2832   8.2450 -21.3782   8.8694  -5.8486
#>   -1.7773   2.4625  -3.9122  -2.7017   0.0538 -18.4840  -6.9113  10.7328
#>   -3.5684  -2.7977   1.4231   9.4767  -5.1381  14.3367 -21.5574   0.1931
#>    6.3921  -7.7233   5.8988  -5.3816   9.9452  -5.1910  -6.7576   8.4177
#>   -3.4592   0.3654  -0.2393  -4.6396  -5.2967  -3.3108  -8.3556   2.4349
#>    3.3613  -1.1057   2.7994  -2.1048  -0.1011   2.4599 -17.4933 -11.7592
#>    0.4766  -3.2943  -0.3698  -1.8658  -9.4721 -19.7548  10.7719  -8.6602
#>    4.2663  -6.8450  -4.4745   2.7198 -11.3084   6.9812  -2.0947   0.7545
#>    0.8951  -3.5025  11.2053  -1.8385  -7.5877   3.6083   0.3264 -13.5170
#>   -3.5710   5.1281  -0.9603 -10.7517   4.9555  -5.7363  10.0621 -14.9625
#>    8.5768   5.3510   0.0473   2.0731   2.7863  14.7675  -8.3409  -2.5829
#>   -0.6405   3.7741   1.7174   2.1088  12.9921   4.7044 -13.0797  21.5564
#>   -5.5350   4.6023   3.3371  13.9133 -10.1255  10.2307 -16.2516 -11.0651
#>   -4.0307  -4.3944   9.5235  -1.1854  -7.8096  17.5948   8.7364  17.7955
#>    6.0037  -5.4486  -1.0751  -2.3095  -5.6568 -13.8471  12.5653   3.3867
#>    0.9275   0.4495  -7.6345   6.5766   2.4717  -6.3577  -5.5690   0.0838
#>    1.0183  -0.2314   0.1864  -7.3133  15.7670  -2.3182  14.9304  13.8206
#>    3.1376  -5.1383  -2.1301  -3.5807   9.1248  18.5492   6.6417 -15.0066
#>    5.7975  -3.7254   7.6468   8.2287  13.3673  10.6231  12.2382 -14.2414
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]