Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -7.3106  -0.4610  -4.7033  -7.6429   8.7402  15.1950  -6.5458  21.9319
#>   -0.8532  -4.1527  -7.3687  -5.5622  -5.4141   1.1717  -2.8204   4.0608
#>   -2.0782  -1.2242   0.9435   5.4283   2.0150   7.6492  -8.5047   9.5300
#>    3.3804   0.1901  -0.0409  -8.2922  -6.0467   9.5296  -6.8060   1.5016
#>   -1.6172  -0.0587   2.6989   8.2870   1.8990  10.8816  -3.1680 -20.6097
#>   -5.9941   4.8334  -8.7956   5.2046  -9.0254  -6.2253   2.3649   8.0859
#>    2.5009   5.4282  -4.6832   9.5208  -3.0278   9.8432   2.0564 -12.0559
#>   -0.4362  -5.3910  -2.6438  -1.6531  -9.2846  -6.3974  -5.1136  13.6939
#>    5.2211  -5.2570   3.7978   1.0088 -21.9180   9.3626   0.9905  -2.6643
#>    7.5091  -4.7673   5.4660 -10.4830 -14.2816  10.8633  -4.5133   2.9388
#>   -4.6496   3.9599   8.2624 -11.8139  16.4332   4.0719  -2.9143  -3.3701
#>    5.5910   6.3742 -10.7537   8.7515   1.5773   4.9372   4.5482  -0.6570
#>    8.9702   3.9721 -12.3725  -1.0572  -1.2665   3.1206   0.8692  -5.7656
#>    7.6692  -4.7151   3.9089  10.0255 -15.0771   3.9222  -2.6106  -4.2545
#>   -5.1735  -5.2065   1.2352   1.7800  -6.6777  14.5841   6.8727  -7.1092
#>   -1.3233  -5.0832  -2.7392  -5.9056   2.8638  -6.6085 -10.0465   3.2296
#>   -3.6957   0.2089   1.9978 -12.5499  -0.8040  -7.3056  -8.7752  19.4814
#>    3.1781   4.7238   6.2456   7.1993 -14.2640 -11.6152  -1.1609   9.8340
#>   -1.6427   4.1660   0.3353  -5.9725   6.6133  -3.1559  -8.0803   7.8872
#>    8.9036  -1.9592  -2.2922 -13.5889  12.1726   2.3078   6.2063  -3.2128
#>    2.6999  -8.4087   4.1266  -2.0208  -6.2850   9.2409  -6.2937   0.9224
#>    3.8466  -3.3499   2.3058   7.5812  -1.3468   7.6777   9.6540  -6.5787
#>    1.5808  -2.2726   2.9409  -8.5640   1.8725   4.5815 -11.0085  13.1409
#>    1.4707  10.2798  -5.2919   7.8878 -17.9565  -5.1337  -6.3306   3.6505
#>   -2.8657   0.2601   6.4001   1.8847  -1.3400   3.0369  -4.2477   8.8686
#>    2.6287  -2.2392  -2.2949  -5.4386  -4.2952  -8.0198 -19.6316   5.6494
#>   -0.5789   1.7234   0.9003 -10.3349   3.4699 -10.8668 -11.9265  16.6447
#>    2.9997  -5.4204  -5.1356  12.4279  -2.6287   1.1163   6.8337 -13.5985
#>    9.4285 -10.8147   1.7952  -1.7065   6.9117   5.2015  22.1181  -4.8371
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]