Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.3854   4.0605   2.6092  -7.5107  -3.5072 -13.2902  -5.8376  -2.5036
#>    6.2752   5.3703  -8.7003   4.4899  -7.0122  -1.8485   9.3786  -0.0282
#>   -2.0421   4.7939   8.2109   3.1436  -4.8922 -12.1591   3.5775  -0.9650
#>    2.4131   1.9616   2.9301  -2.2271 -14.2444  -5.2426  10.2866  -4.4738
#>   -0.6274   1.1308   1.6537  21.4150   4.8937  -5.0842  11.6069  -0.5253
#>    1.7462  13.5773   2.7830   4.3857   7.9440 -10.0400   2.8088  -1.4426
#>    4.6636  -2.8870 -13.5024   9.6159   5.3605  -3.3262  17.9829  -5.0171
#>   -6.3440   2.8888 -12.1036   7.4695  -0.1920   8.5471   2.4986   0.6234
#>   -3.4167  -9.5803  12.9874  -6.7163  10.2770  -3.2718   1.0825  -6.3130
#>    1.5140   1.2291  12.9482  16.2996  12.7175   2.6251  -2.2271  -8.5858
#>    0.8426   5.4524  14.3257  14.3597   8.6324  -9.5009  -3.3864   2.8085
#>    1.9225   7.2875  -4.3488   1.8059  -3.0599 -10.0045  -1.8368   2.5847
#>    0.7364   2.5947  -3.4403   5.1562  -1.3948  -2.6911  -7.1656   8.5713
#>   -2.0360  -2.3181   6.5904  -1.1158  -1.8095  17.0864   4.2294  11.7323
#>   -2.0568 -12.4782  -6.9439  -5.5109 -18.0618   6.3722  16.3761   7.5172
#>   -2.7137  -3.1583   1.1402  -3.9024   4.0540   2.4089  14.5710  -8.1386
#>   -0.0392  -0.8874   8.9106   4.6033   3.0890  11.1206   2.0823   5.8826
#>   -0.9914  -1.5002  -5.6391   1.5676  -0.1772   5.5227  -1.6010   4.4761
#>    2.0946  -2.1132  -6.9960  -9.3121   0.5019   3.8096  -3.8729   8.5361
#>   -6.6675   2.9183  18.7688  -2.6221  -5.7250 -10.0844 -17.6697   3.3959
#>    6.7573  -1.6437  -2.3026   0.0159  -0.7178   2.0287 -14.3911   8.0814
#>    4.4475  -9.1442  -2.4342   5.4868  17.4149  -3.2796  10.5417  -2.6966
#>    1.9502  -0.4939   4.9881 -11.1700 -12.8852  -3.3846   2.0790  -7.9310
#>   -4.0029  -5.6472  -4.5148  -8.1974   3.8486  -3.1410 -13.9316   5.1537
#>    2.8976   1.9508   1.6538   2.3306   1.8485  -9.0188  -5.6464  -2.5932
#>   -7.8989  -7.9292  -3.3941  -8.5549   8.0725   1.7404 -20.2085 -12.6783
#>    3.6424   4.8223   4.7794   2.1899   9.9453   1.6169   4.2087  -2.2596
#>    7.2475   0.2194  -0.6088   0.9401  -8.7628 -11.8543  -3.9840  -4.4968
#>    9.2541   4.6107   8.3379  13.4325   1.6763  -5.3736   6.6128  23.9814
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]