Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  4.3787e+00  5.4327e+00  1.1606e+01  8.7054e+00 -5.1792e+00 -9.4609e-01
#>  -6.3415e+00  1.9131e+00 -5.8820e+00  6.2941e+00  8.5524e+00  8.5581e+00
#>   1.9419e+00  5.7455e+00 -4.9094e+00  5.4367e+00  9.9761e+00  2.8589e+00
#>  -4.0702e+00 -2.6408e+00 -1.8930e+00  1.1163e+01 -5.0533e+00 -2.5793e+01
#>   8.8065e+00 -1.3453e+01 -5.4858e+00 -1.5626e+01 -6.3268e+00  1.1318e+01
#>   4.8812e+00  6.0748e+00 -1.6551e+00  1.7636e+00  4.1255e+00  8.2965e+00
#>   4.7515e+00 -5.2365e+00 -8.5680e+00 -5.0484e+00 -5.6363e-01  5.6351e+00
#>  -3.7313e+00  7.1505e-01 -3.7328e+00  2.5697e+00 -2.0078e+00  1.2310e+01
#>  -1.4350e+00 -6.0564e+00 -4.4675e+00 -8.4813e+00 -4.3616e+00  7.7436e+00
#>   3.4569e+00  2.2377e+01  6.5848e-01 -1.3907e+01  6.5336e+00 -1.6113e+01
#>   4.7035e-01 -6.5569e+00  2.6716e-01 -8.8895e+00  1.1162e+01  6.3563e+00
#>  -8.5916e+00  6.8292e-01  6.6782e+00 -2.0376e+01  5.1052e+00  2.2968e+00
#>  -3.3653e+00  9.0146e+00 -3.8344e+00  1.1825e+00  5.6884e+00  2.0531e+01
#>   1.5351e+00 -6.7317e+00  1.7165e+01 -3.3459e+00 -5.6619e+00  1.9369e+00
#>   7.9569e-01 -2.8745e+00  9.6786e+00  1.9100e+01  2.6596e-01  5.1515e+00
#>  -2.5254e+00 -1.0088e+00  6.7740e+00 -2.7019e+00 -3.0651e+00  1.7116e+01
#>   1.3089e+00 -3.0917e+00 -9.8509e-01  1.5425e+00 -3.2869e-01  5.6889e+00
#>   4.1197e+00  2.8519e+00 -1.2346e+01  1.6896e+00 -1.6579e+00 -1.8745e+00
#>   4.3207e-01 -4.8242e+00 -4.0499e+00 -9.1290e+00  7.0664e+00  1.1969e+01
#>   2.5248e+00  2.0511e+00  3.8594e+00 -9.3571e+00  4.7500e+00 -1.1038e+01
#>  -4.9330e+00 -4.4742e+00 -4.3901e+00 -1.0956e+01 -1.3986e+00 -6.2840e+00
#>  -6.4236e+00 -8.4028e+00  9.5905e+00 -1.4650e+01  4.4810e+00  1.4386e+00
#>   1.7769e+00  7.9861e-01 -1.9396e+00  7.9887e+00 -2.6267e+00 -6.9383e+00
#>  -4.3963e+00  4.0397e+00  7.8207e+00  3.2191e+00  4.0039e+00  1.3717e+00
#>   3.7805e+00  1.5971e+00  1.6282e+00  6.3334e-02  4.7413e+00 -7.3664e-01
#>   6.0521e+00  1.2462e+00  4.5857e+00  7.1756e+00  1.7437e+00  1.8700e+01
#>   1.6095e+00  2.3357e+00 -1.7958e+00 -8.8527e+00  5.3082e+00  6.6395e+00
#>   5.2093e+00 -2.3623e+00  9.2423e+00  2.0400e+01  6.9068e+00  4.6434e+00
#>   3.9018e+00 -1.6464e+00 -1.3899e+01  1.5016e+01 -1.5874e+00 -5.2777e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]