Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -3.3022e+00  2.6165e+00  6.6420e-01  1.5516e+01 -3.1969e-02  9.7983e-01
#>   5.7152e+00 -4.4268e+00 -9.2472e+00  2.0005e+01 -7.8061e-01 -5.0885e+00
#>   1.8950e+00 -5.6772e+00 -1.1335e+00  6.2996e+00 -2.9074e+00  4.2991e-01
#>   1.6300e+00  3.9305e+00  4.9314e+00 -2.8041e-02 -2.6692e+00  2.8712e+00
#>  -3.9994e-01  1.1126e+00  3.3731e+00  3.0055e+00  3.2497e+00 -6.0785e+00
#>   1.1627e+00 -9.6512e+00 -2.3807e+00 -2.2724e+00 -2.3733e+00  7.6403e-01
#>   7.3658e-01  3.0814e+00  7.4240e+00 -5.8574e+00 -1.6081e+01 -1.6479e+00
#>   2.8044e+00  9.8602e+00  4.8417e+00  4.8215e-02 -2.0838e+00  2.6613e+00
#>   1.9153e-01 -4.4462e+00  1.1578e+00 -6.4918e+00  1.2274e+01  4.2213e-01
#>  -7.0114e+00 -6.3886e+00  1.3247e+01 -1.0664e+01  3.0986e+00  1.9476e+00
#>   8.0956e-01 -6.5787e+00  5.0560e+00 -4.8809e+00  3.5571e+00 -1.0171e+00
#>   1.7309e-01  2.4220e+00  4.3167e+00  2.3469e+00 -5.4695e+00 -3.0518e+00
#>  -2.6766e+00  1.9773e+00 -1.2293e+00 -1.3348e+00 -1.3324e+01 -1.6468e+01
#>   2.2509e+00  4.2020e+00  1.4450e-01  1.1577e+01 -8.5960e+00  2.5525e+00
#>  -1.6889e+00  4.2297e+00 -5.3866e+00 -9.5325e+00 -5.0040e+00  1.1619e+00
#>   1.4893e+00  2.1871e+00  9.0348e-01  9.3735e+00 -1.0362e+01  1.1437e+01
#>  -2.2720e+00 -1.8302e+00  4.9754e+00  3.6586e+00  6.9248e+00 -2.7575e+01
#>  -3.6859e+00  7.2915e-01  3.3375e+00 -5.5262e+00  1.4209e+01  8.8956e+00
#>  -9.0200e-01  3.8231e+00 -2.1662e-01  1.3115e+00  8.0924e-01  5.4239e+00
#>   7.6760e+00  2.2301e+00 -2.6176e+00  2.2741e+00 -6.4941e+00  1.5622e+00
#>   3.2242e-01  1.0041e+01 -1.2422e+01 -6.1486e+00 -2.9464e+00 -5.0572e+00
#>   5.0351e+00 -4.6980e+00 -5.7113e+00 -5.3465e+00 -9.0946e+00  8.6780e-01
#>  -6.7897e+00  5.9060e+00  3.9313e+00  3.9744e+00  9.8440e+00 -1.5659e+00
#>  -4.7420e+00  4.0920e-01  5.3746e+00 -8.0431e+00  5.4999e+00  1.2874e+01
#>  -6.2298e+00  7.2065e+00  6.1496e+00 -6.0684e+00  3.4760e+00 -3.4340e+00
#>  -3.1890e+00 -1.9099e+00  9.7531e+00  6.7009e+00 -5.7481e+00 -1.5809e+00
#>  -1.4285e+00  3.8691e+00  5.7148e+00  6.9438e+00  1.3401e+00 -8.1290e+00
#>   2.5297e+00 -4.5300e-05  4.9108e+00 -1.0668e+01  5.1974e+00  2.2742e+00
#>   1.9452e+00 -6.8699e+00 -9.1499e+00  1.6425e+00 -1.2386e+01  5.3117e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]