Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.7118  -8.9964   9.3219  -0.5564  -7.2135  13.5453  11.9109 -11.1107
#>   -9.2528   5.3217  -4.5195  -3.0534   2.8512 -14.7239  -6.3183 -10.7822
#>   -5.6344   3.1621   1.7118  11.1003  16.8259  10.2988  -0.9521   5.9340
#>   -1.8525   0.7704  -4.1343  -2.2158 -15.5819 -12.5315   4.7872   4.3192
#>    0.1541  15.3781  14.5941  -1.4536   0.8289  10.0110  10.0660  -2.0917
#>    3.8165   7.6666  -4.9484 -16.8344   3.4838  11.1285   4.8758   1.4582
#>   -0.5078  -0.3781   4.4112  -0.1791  11.0184   0.5833   8.5735  -6.2380
#>   -1.6065   0.4467  -0.9730  -1.7176   2.7749   9.2794  -9.5734   3.2870
#>    5.8621 -11.9224   0.0285  15.8147  -0.8104   8.5851  -9.7686  -2.1034
#>    5.2970  -7.1224   2.9436   8.8525   8.6265 -16.3642 -18.0349  -4.1870
#>    3.0181 -17.2284  -0.4494  17.2609  -1.0383 -19.2469  12.3642  -9.8111
#>    2.6622   0.6141 -11.8549   7.7994  -5.0168 -16.2010   5.1448  14.0812
#>   -5.3456   1.8860  -4.7246  -8.4667   8.6556   3.1856   9.6151   5.1781
#>    1.9482   1.3096 -13.3742 -11.1149 -11.7821   7.3021 -11.6727  -5.1679
#>   -6.4591   9.2479  -4.4072  -4.6665   5.1705   2.7141  -1.6902  11.7043
#>    3.2724 -15.7130  15.4191   6.0613  -3.6882   7.0697  -5.8291   0.6819
#>   -0.1117  -3.4258   2.3955  12.7584   2.8108  11.9407   5.9166   4.6668
#>   -2.8276   1.7866   2.8007  10.0424 -15.9661  -6.2253  -1.1789  -0.6139
#>   -7.2391  -7.3002   8.2192  -4.6093   6.6921  -0.7069   2.3826 -11.2901
#>   -0.7116   4.7317   1.9027  -0.9118  -4.5538  -1.1300  -1.2870 -14.0126
#>   -1.3290  -2.5808  13.7263  -3.9387   7.0996  -7.1606   0.6745  -2.1389
#>    2.2780  -3.9044  -2.9945  -0.6770   7.5248  -2.7557   3.2035  -5.0747
#>    6.8374   1.4933 -11.9957  -7.0147 -11.3972  -0.8300  -4.4271  10.1217
#>   -1.6824  -2.9623   7.7868 -18.8423   8.5845  -4.8876  13.8579  10.2705
#>   -2.5902   6.5935  23.7773  -8.2538 -15.9391 -13.8985  -7.2517  -2.6537
#>    2.0002   5.0033   0.8899  13.1341   0.0274   1.9275   1.6034 -13.5794
#>   -3.7410  -1.7408  -5.2692   0.8601   7.8146 -10.1747  11.0711   1.9217
#>   10.0803   0.3754   3.3139   2.4871 -10.6417  -2.1276  -1.1723  -6.1042
#>   11.0186   0.9802 -18.3512 -11.5617  -2.6583  23.1547  14.4453   0.0101
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]