Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.1222  -5.4486   7.2633  11.3367  -0.7066  -7.5129  -2.5550  -2.3037
#>    1.1049 -11.6712  -1.4263  19.2739  -2.4047   9.2700   3.2903  -2.1753
#>    0.2270   3.8546 -11.7057  -2.5193  -4.7407  -3.6497  10.8778  -8.7295
#>   -4.5926  -2.4497 -21.0203  -7.5067   0.8255   1.6252  -8.2317  -9.1015
#>   -1.6336   7.5147  -1.4509 -11.9998   9.8035   1.4634   5.2402  20.7168
#>   -6.8147  -3.4019   9.6589   9.2282   5.0617  14.5221   2.7119   2.1456
#>    7.8990  -4.0032   5.6035  11.4265   5.2796   0.0367 -14.9254   1.9565
#>   -7.6337   5.0451   9.0863  -2.4556 -22.6308 -15.1475   8.9885  -7.6004
#>   -0.6400   4.6194  11.6725   3.4513  -1.1060 -14.6762  -2.0422  15.2965
#>   -0.0355  -5.5114  -3.5756   3.1303   9.9892  -2.6016   7.0588  -4.5740
#>    1.7430  -7.3117  -2.6017   7.5779  -8.6797   2.9843  -0.3459  -2.7696
#>   -2.4468  -8.2330  -1.1111   7.2897  -0.8333  -5.1176  10.2867 -16.4666
#>   -0.5694   9.3861  10.8545   1.5810 -15.9735 -11.0388   6.1639   3.1335
#>   -5.8092 -11.7664   4.7074   5.1425   4.6644 -13.9152 -11.6685   0.5808
#>    5.6462   1.2163   0.9277  -0.0944  19.0261 -10.3951  11.8321 -10.7024
#>    1.8310   4.3073   2.7645 -20.0861   3.8596 -10.0840 -14.2402   5.7745
#>    2.7310  -2.1179  -8.0965  -1.4818   5.5919   1.6979  11.7188   4.3893
#>   -0.5601  -4.5799  13.1112   6.3013  -0.8812  -1.5572  -1.1828   1.1050
#>    0.2875  -4.9229   4.7585   0.9537 -16.8656  -6.5139 -16.5307   4.6824
#>    0.7466   0.2649  -9.4189  -7.5092   9.3359  -2.2484   0.6703 -15.4507
#>  -12.2948  13.4318  -2.2880 -14.3878  21.5949   1.4125  -3.1950   6.6372
#>    4.8411   0.6713  -4.0318  -7.3650  -5.5390  -7.4694 -12.2712   1.5075
#>    0.3039  -0.1851  -0.0877 -11.4101   5.2430  -4.1657   9.7696 -10.4608
#>   -2.0572  15.1112  -5.7256  -1.0771   7.2850   7.1663  -1.5657  -4.1502
#>    1.4187   0.5693   3.3325 -10.1046  -8.8140 -12.7380   6.1884 -11.8620
#>    1.7053  -3.6630 -20.0785  -6.6436 -22.4304   1.6951  21.5032 -21.4564
#>   -2.1100   3.3554  12.1688  10.9437  -3.4092  -1.6302  14.3686   0.7109
#>   -3.3155  -3.1158  13.2419  -1.0228   8.0589  12.8834  -5.1307   4.3118
#>    0.6990   1.8644   2.7455  -4.3046  -9.9062 -10.2970  -8.9674  18.9196
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]