Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.9998  -6.5477  13.9528   9.8868   7.7534  -0.3696 -13.6729  11.8097
#>    4.6908  10.5226  -3.4257   0.7841  -4.5433  -0.2350  -2.8775  -3.0741
#>    3.6050  -6.8092   0.4119  18.2815  13.6924 -17.0933   4.6874   6.5449
#>   -1.8907   7.9875   0.7440  -5.4101  -3.0243   3.7560   4.7827   5.0181
#>    2.5163   7.3509   2.1008   8.0190   5.4385   0.6389  -0.9474  -1.3100
#>   -4.9332  -6.7909  -4.7102  -3.1487  -1.3700  -8.7599  -4.9827  14.5740
#>    1.2629  -1.0223  10.1007   0.7744   4.8901  -4.2751   0.2163  -0.7365
#>   -8.0232   2.9351  -2.3465  -4.6006  -1.2983   2.5331  -6.0859  -8.0989
#>    0.9087   4.3857  -6.5489 -19.4752   8.0951  12.1559  13.5094   6.9974
#>    1.9606  -2.1891  -2.4691  -9.4395   1.4261  -0.4516  -7.6738   3.2115
#>    2.6585   0.8101   6.8073  -4.5371   5.3780   3.5464   9.1067  14.1215
#>    6.8516  -6.1839  -4.7957  -7.9348  -7.4111 -11.4674   1.6115   9.7141
#>    0.6147   8.8971   5.7957   6.5988  -0.2715 -18.4819   0.4815   0.3597
#>   -1.2568  -0.4924 -11.8895   8.9526  21.1243  -8.4383  -8.6565  17.7103
#>    2.3979  11.5278   0.4399  -2.8978  10.0500  -0.1760   2.3308  -7.6613
#>   -1.5282   3.4834  -3.4036  -0.9245  -5.3371  11.5379  -3.7916   7.8611
#>   -4.4416   7.6587   0.4796 -11.0805   7.5640  -2.5914  -8.4839  -0.8025
#>   -5.7319   0.0374   7.5358   0.7486  -5.7952 -21.4179 -19.5100  -7.3978
#>    6.0816  -0.5011   2.7554   0.8352 -14.0759  -6.8995  -5.9869   8.7820
#>   -1.2777  -0.2300  -2.5704 -17.9806 -24.1033   5.9010 -12.7177 -17.0192
#>    0.2233   0.9853   5.1328  -2.4025   5.7548 -15.9731  -4.1422  -8.4812
#>   -4.7457   3.7227  -1.4841   6.0990   5.7879  -5.8002   5.6304   0.0509
#>    4.6902  -8.8986   3.6852  -2.9334  -1.5350  14.8437  -6.8987  -0.7912
#>    0.0835  -0.4865  -8.0517 -22.3858   1.3133   3.1928  -0.2214  -8.3891
#>   -0.2852  -4.2740  -6.6811   0.0128 -11.5420  10.2900  -0.6745  -8.0989
#>   -0.8496   4.5299  -0.8760   8.5970   6.8358   9.5024   2.6338   1.6452
#>    4.9016   0.7758   0.5276  -8.7958  -3.5791  -4.3830   9.1531  12.8765
#>   -0.2930  -1.5542   6.0205  -6.6197 -16.0742  -6.8010  -7.2781   6.9355
#>    1.5136  11.7653   5.3117   9.3849   1.6369  -9.4203  -9.5110   1.5041
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]