Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.7322  -7.2974  -4.4590   0.7428  -9.4571  -2.5948  -1.3927  -8.6694
#>   -0.7859  -6.0713  -4.4713   1.0666  13.6333  -3.9293   9.6050  -2.4316
#>    2.8941   2.6818  -2.9872   8.5088   7.5633 -11.0718   1.3690  12.1845
#>   -1.6419  -4.1953   0.7662   8.2677   7.0667   8.7502  17.2932  -9.8404
#>   -2.6648  -7.2944  -4.7588  -4.4150   6.5100   0.4739  11.6600  12.7523
#>    2.1723   9.1907  -1.2885   1.9014 -17.8156   7.5440  25.4154  11.7736
#>    4.1733   4.6096  13.0876  -9.5356   3.4294  -8.7730  10.1644 -14.4631
#>   -0.0656 -10.8829   1.0563  -0.8801  -1.3110  14.6729  -1.6392  -8.9219
#>   -1.4167   2.7224  -2.0914  -1.2642   7.0885   3.8600  -1.3786  -6.2784
#>    1.5536 -12.4498   7.0038  -2.3915  -6.2942  -5.4538  19.4580  -7.3315
#>    3.5581   4.8065  -1.1288   4.7433  -8.1135  -2.1545  -0.8185  -3.1584
#>    1.4699   1.4805   7.4189  -1.1814  -6.3836   1.2016  11.2856  -7.5401
#>   -2.3628   8.1299  -6.0979  -1.2912   8.7419   5.0470   2.0521   2.2154
#>    0.1019   0.6051  14.5055 -11.5754   6.8237 -10.7215   9.1222   9.0923
#>    4.8043  -7.8786   5.0379  -7.3304  11.4224  -9.2436  -5.5077   7.9933
#>    2.9188   4.4580   6.0683 -11.5580  -5.9713  11.2154  18.0074  -5.1108
#>   -0.7265   1.8772   0.2097  -2.6194  -3.8067  13.7727  -8.8926   4.2694
#>   -0.1094  -5.1905   1.1261   3.6791  -6.7643  -8.4307  12.8281   8.4641
#>   -2.0830  -1.6228  -7.8450  13.1146   9.1031 -12.4419   3.4959  -7.8593
#>   -5.7280  -4.7057  -1.2586  -3.3872  -8.4106   5.4395   9.6756  10.7280
#>   -2.7143   1.2538   4.4405  -7.8312  -5.4987  -0.4699   3.7534   6.9985
#>   -1.5433  -9.4743   7.8347  -7.3382 -14.5286  11.9636   5.9747  -1.2699
#>    3.0691  -4.8899   5.8841   2.4252 -10.2630  -6.0420   1.8649 -10.6841
#>    1.8828   7.0520   2.0558  14.2881  -3.9991 -19.3342   7.3384   0.7462
#>    7.1091   4.6440   5.7643   0.9587  -7.0511  -3.6311   1.9882  10.5506
#>   -1.9701   2.3204  -0.6040  -2.3789 -23.0143  -0.3606 -18.6548  16.2919
#>   -1.8938  12.4914  -7.7991   4.4974   1.0664   1.6096  -5.7728  15.2738
#>   -2.1721   3.4871  10.5756 -14.4790   8.3504  16.9059  -5.3754  -0.7694
#>    0.5197 -10.7416  -5.3533  -6.4701   0.6537   6.8183  12.6312  -4.8186
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]