Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -4.8730e+00  4.7144e+00 -2.3117e+00  2.2992e-01 -1.0033e+01 -1.1192e+01
#>   2.5941e+00  1.6443e+00 -1.7006e+00 -3.1548e+00  1.1195e+01 -4.8058e+00
#>  -5.8910e+00  2.1196e+00 -2.3706e+00 -1.0519e+01 -1.7014e+00 -7.3250e+00
#>  -2.8502e+00  3.5153e-01 -1.1059e+01  2.3682e+00  7.6513e+00 -4.1446e-01
#>  -2.7890e+00  1.5020e-01  1.3857e+01 -1.4057e+01 -1.4487e+00  2.2806e+00
#>   7.2235e+00 -2.4472e+00 -5.3875e+00 -5.1582e+00  4.9037e+00  8.2627e+00
#>  -1.6541e+00 -1.7581e+00  3.0510e+00 -2.5223e+00 -1.1292e+01  7.4508e-01
#>   1.3349e+00 -8.7178e-01 -3.9173e+00  7.7994e+00 -5.2549e-01 -9.8688e+00
#>  -4.8321e+00  4.5818e+00  1.6677e+00  6.8650e+00  3.8342e+00 -1.8017e+00
#>  -1.5861e-01 -1.4499e+00  5.2906e+00  4.1680e-01 -6.6700e+00 -1.1750e-01
#>  -2.7359e-01  1.3487e+00  5.1425e+00 -6.1243e+00  9.4078e-01  9.0915e+00
#>  -3.6741e+00 -1.8984e+00 -5.7932e+00 -5.6399e+00  1.0565e+01 -8.6918e+00
#>   5.8719e+00 -7.7216e+00  1.2867e-01  1.3522e+00  8.1807e-01  8.9480e+00
#>   1.5693e+00  2.1457e+00  5.2436e+00 -1.0419e+01  4.1283e+00 -8.3191e-02
#>  -5.6722e+00  5.0688e+00 -1.5361e-01  3.7354e+00  8.5997e-01  1.5671e+01
#>  -3.2571e+00 -9.1009e+00  1.0099e+01  6.4411e+00  8.2171e-01  2.0834e+00
#>   4.4435e-01  2.9568e+00  4.5476e+00 -9.0577e+00 -3.2558e+00 -4.4935e+00
#>   1.7777e+00  2.6863e+00  1.5264e+00 -2.1159e+00 -2.7905e+00 -4.7954e-01
#>   2.7409e-01  9.9658e-01 -1.7825e+00  4.3340e+00 -4.6216e+00 -6.2509e+00
#>  -3.6393e-01 -3.2049e+00 -3.0196e+00 -9.6815e-01 -6.5626e-01  9.9816e+00
#>   3.8720e-01  6.5430e+00 -1.8441e+00 -5.2576e+00  3.2002e+00 -4.3290e+00
#>   9.0745e-01  1.1153e-01  3.0420e+00  6.2075e+00  1.1284e+01  3.5742e+00
#>   3.1579e-01  6.0095e+00 -1.3809e+00 -2.2897e+00 -4.4049e+00 -1.3696e+00
#>  -6.4769e-01 -5.8033e+00 -5.7902e-01 -4.1393e-01  7.5510e+00 -4.2001e+00
#>  -3.5596e+00 -8.9236e+00  5.2550e+00  3.1106e+00  8.1544e+00 -1.3634e-01
#>  -4.0053e+00  4.1994e+00  1.8587e+00  5.1971e+00 -1.3432e+01  7.7787e-01
#>  -4.5144e+00  3.1390e+00  7.6420e+00  9.4894e-01  4.9226e+00  1.6994e+01
#>  -1.1126e+00  1.5032e+00 -6.2700e-01 -5.7691e+00  6.5679e-01  5.4799e+00
#>  -2.6061e+00 -7.3668e+00  1.8084e+00 -1.1680e-01  1.3367e+00 -1.4254e+01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]