Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.6468   3.9243 -13.2213  16.7991 -24.1294  19.8588  22.3439 -15.0503
#>   -1.3519   7.8718   7.9608 -13.1395   3.1858  15.1393 -12.7564  -1.9935
#>   -0.5137   4.8682   8.0694   3.8156  -5.6112  10.4141  -1.5149   8.2850
#>    0.4397   6.2588 -10.7781  -4.5205   3.5962   2.0486  12.6823 -14.9501
#>   -5.6103  -8.1864   6.9209  -2.2391  -3.4078   6.8275   1.3508  -7.5091
#>    2.6352 -10.3915  -3.9605  -0.0862 -25.5401 -18.8536   4.7775  -7.9937
#>    4.4831  -0.2244   8.1950   8.9697  -8.4163   4.5950  -0.2130  -6.4934
#>   -0.2286  -4.1280   6.1501   9.5755   2.4431  -6.1918  -5.7236  -3.2768
#>    0.0819   4.6372   3.3003  -5.1310  -5.7576 -12.1548  -5.1150   0.8308
#>    0.7352  -3.0385  11.1368  -3.4963  12.8064 -21.9484 -15.8961   9.6473
#>    0.2349   5.9535   5.1350  -5.4402  13.6027  -1.3196  -3.6131  22.1304
#>    2.4023   0.4497  -2.3792   8.5501  -4.2540  -9.9833   8.3947   1.1608
#>    1.2907  -3.5542  -1.3944 -16.5227  22.9701  13.7866 -17.3123   1.1753
#>   -5.8432  -5.9766  -3.7574  -3.6187   7.1735   5.8945  -9.1497   2.4915
#>    2.2222 -10.2757  -5.4079  -0.2531  10.0265  -1.0054  10.0633  22.0880
#>   -1.2755  -2.4352   0.4808   2.5882  -9.5818 -12.0250   6.1864   1.5529
#>   -1.0782 -10.2271  -2.8395  16.1301  -2.1499  -0.9267   4.7660  -7.0072
#>    6.8478   0.5709  10.2586   6.7352  -5.8979   0.4806  -0.0110  -8.0678
#>   -5.0488   0.5993  -1.6806  10.6423  -2.9808  -0.4496  -2.5613   3.8187
#>    0.9046   7.3943  -0.8032  -6.8459   9.7090  -3.4071  -7.7073  14.2601
#>    2.7042   5.7376   2.7225   3.9296   6.9634  -0.1143   8.3694  -3.1764
#>   -0.6475  -2.4878  -0.2758  -0.9828 -10.8199  -2.1731   0.7606  -4.5758
#>    5.8344   4.6115   0.7985 -12.4002   6.4071  19.8737  10.7113   0.5568
#>   -0.9887  -4.7737   4.1255   0.5020   8.5736  15.2775 -11.1597   4.0921
#>   -2.2699  -7.4462  -4.5329  10.9930   5.0434  -1.1165   3.5150  -7.2903
#>   -1.5349   5.5053   9.0535  -7.3091  -4.7176  -2.5532   9.1938  -5.0839
#>   -0.3786  -5.9113   0.4431  14.9552  -4.7897  -1.8035  -3.8073  -2.1094
#>    4.7222 -10.0281  13.5130  18.9236  -2.9412   5.4265   8.1387  -5.8685
#>   -0.5742  -5.3046   6.0370   1.0182  -3.6745   2.1683  10.1980   5.3282
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]