Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.2204  -7.9417  -2.1131  10.5518  -1.3744   6.3677  -4.0676  -8.3950
#>   -3.1571   1.3070   8.8680   8.4880   3.0009  -7.4361  -2.6198  -6.8878
#>    3.1001  -3.4728  -2.4978   3.5462   6.7748  -3.9296  -4.1638   6.1255
#>   -2.6534  10.8842   3.6607  -3.6120  -4.0002  -7.5365   9.4299  19.2406
#>   -0.0713  -5.8547  10.2349  -5.8379   0.9194   6.1737   5.4703  -4.1014
#>   -0.7283   3.1369  -7.1914   2.0305   7.6965  -5.3969  -4.4720   4.5913
#>   -3.5930  -1.8463  -8.3993   3.2868   7.0131  -4.4316  -6.4816   0.2510
#>    0.3346  -0.4794   4.8879  -1.1231   4.6889   2.6526  -0.0265  -8.4203
#>   -1.5554   4.2325  -2.0063   4.7472  15.0558 -10.1606  -1.1503  -3.8528
#>    0.2759  -3.1255   2.1473  -0.4764  -4.1435 -12.1488   5.3734   3.4103
#>    0.1867   4.0412   0.9623   0.1999  -0.4423  -4.0111   8.8624   3.9210
#>    3.0697  -5.5143  -2.4931  -2.5820  -3.2105   0.0511 -11.6968   1.2614
#>    0.8474  -0.1987 -10.7537  -6.4094  12.7619  11.0048 -13.4024  -1.7844
#>    0.1137   0.5254   5.1498   4.6826  -7.1421 -10.3179  -8.6732  -3.2845
#>    0.1554   2.1473  -2.6701  -7.1717   7.9441 -13.9724  18.2807  -5.5683
#>   -1.8801  18.8454  -3.2451   9.7359  -3.8447   7.4490 -14.6456  -1.0630
#>    2.9151 -11.4855  -8.2246  14.3100  12.9537  -3.4432   1.4322   7.7472
#>    2.6505  -4.3001  -9.0467 -17.4403  10.2788   4.4977  -1.3523  -1.9880
#>   -0.9135   1.0207  -4.0653  -3.8313  -4.3434 -10.8927   3.6370  -0.8991
#>    0.2258  -6.2240   2.5547   5.5010  -9.2543   7.1128 -10.2785 -14.2051
#>   -4.8800  -0.8872   8.3377 -21.5587 -12.1361  -9.1295 -10.6358  -2.9974
#>   -3.6523  -5.0106  -5.5534   8.9988  -1.4298   0.7248  -1.7000   3.0973
#>    5.9845  -4.1718  -4.2336  -2.5850  18.8871 -12.2306  -4.9356  -2.8990
#>    1.5936  -7.1743  -4.0196   0.2004  17.6340  -3.6541   8.5358  -7.7818
#>   -0.2327   0.8526   6.5054  -0.7430   4.4794   0.5777   6.7966   0.3376
#>   -2.4086   1.2310   3.3151  -0.2862 -16.6893   4.6263  -6.7179   6.0001
#>    5.0495 -13.4504 -13.0339   8.0662   0.9214  -2.9114   7.8740   3.8130
#>   -1.6556 -16.0278  -1.3135  -0.4959   7.7042   5.2776   1.1231  14.5282
#>   -1.2719  -3.6376   1.2590   9.8294  -0.0490 -18.6232  12.0002   7.0659
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]