Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.8888   8.9467   8.4508  -5.9311 -13.3586   0.7778  -3.8868  -5.0629
#>   -6.6587  -1.0274   9.7923   0.3907 -15.3056   3.3187  -0.0768   9.8390
#>    6.4904   2.9006  -0.2438  -5.7461  -7.7968   9.7829  -1.4370  12.4145
#>   -5.5902  -2.1879  -9.1247  -4.2278  -1.8860   0.1672 -14.4110 -12.4069
#>    5.5047   2.2202 -10.3080 -13.5611   2.5965  -4.5777   9.2101  -4.3527
#>   -7.4070   2.6192   7.5689   7.7588  -8.0214  -5.5136  10.2722  -8.1392
#>    1.5999  -3.5977  16.1994  -0.3021   7.3556  -1.9684   7.4836 -10.3204
#>    0.1885   6.2030  -5.0919  11.1331  -0.8414   5.1941  -6.8298  -8.6755
#>    2.9309  -2.6087   7.6690  -7.2715   4.4733   3.0426  -7.7529  -0.2730
#>    2.7659   8.6734  -9.6482  -0.4388 -12.4050 -11.7996   3.1574   5.2909
#>    6.8222  -0.7977  12.6608  -2.5708   0.0272   3.4196  22.2564 -10.0452
#>   10.1662 -13.2170   2.9509  -7.1687   7.9904  -5.9949  20.5783  -0.7673
#>   -9.9701   5.9720   0.6067   9.7075 -14.4777   1.5985  -0.9744  -6.2682
#>    1.8739   4.3883  -2.2135  -5.4524   1.7037   2.9267  19.7503  -2.7806
#>   -4.0710   3.7976  -4.4146 -12.8853  -4.8689  17.6374  -3.3613   2.6218
#>   -6.0181   7.5765  -4.8056 -13.6785   3.3006   2.4838   5.5610  -7.0457
#>   -4.1465   3.4589 -10.3051  -8.5492 -12.1978   6.3886   2.7135   5.6293
#>   -6.8997   0.3344   4.8971  -9.5641 -11.9968 -13.5729   7.0503   4.0081
#>    0.7408   5.5506   4.9726   8.3391 -18.6112  -1.4172  -4.2906  12.6892
#>    2.0710  -0.4447   4.8037   6.2098  -0.8942  12.2368 -10.1579  -2.0934
#>    2.7515  -3.2502   7.6003   0.5885   0.8691  -4.9068  15.8869  -7.4730
#>    7.9047 -11.4026   3.8794   9.2629 -12.3871   0.2359  15.9043  -3.5716
#>    1.5254   4.9538  -4.3943   9.2291  -4.5960 -23.0423  -5.2451   2.3774
#>   -2.5752  -6.3759  -7.3114   7.9352   5.5408  -3.2816   8.2644  -3.9035
#>   -1.7077   6.5352  -1.6821  -0.9688   2.0573  -6.7657   1.6437  -5.2424
#>    9.4966  -0.2277  -9.7141   0.9240   9.1326   7.2720  -9.4410  -6.7757
#>    0.9546  -2.8192  -7.7944  -4.1394   1.9187   2.3440 -10.7446  -6.9024
#>   -2.5247  -6.1437   0.7560   7.9694   0.8405  -2.1161  -1.6735  -5.2300
#>   -0.5687  -3.7463   5.1680   0.3086  -0.3133  -0.3702   3.0327 -15.0728
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]