Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -6.3341   2.7726 -19.5015   6.5292  18.3890  16.6519 -12.3153   8.5059
#>   -3.6244   0.2662   6.8677   1.9725   7.9736  -4.1649 -19.5495   3.8619
#>    1.6318   2.0639  12.6473 -13.8620 -10.3703   3.7272  -8.4758  -4.0649
#>   -3.1007  -3.3835  12.5497  11.2353 -13.6815   4.4780   4.2803   3.9563
#>    1.4438   6.3315   5.0241   9.7363  13.4894   8.4339   2.3855  11.4587
#>   -0.9857   0.7671  -8.1553  -0.9881   4.2799  -5.6123  -0.6646 -18.2653
#>   -0.9423   1.7041 -10.3438   0.4725  -0.8715   6.5590  19.3882   9.3555
#>    3.2103   0.2842 -10.0247  -2.5851  -6.9074 -25.8348 -11.8534  -5.7609
#>    3.0108   1.3241   6.4471 -10.1925 -10.4856   0.9721 -13.4447  -0.8246
#>    0.5372   4.4356  10.6115  -5.6905  10.9131   6.5893   2.5072  13.2452
#>    1.5232  -3.4343   8.3892   1.9707   1.3853  12.9616  -2.4445  -6.0442
#>   -3.0346   4.1571  -4.2184   5.6038 -10.1179  -3.5684  -0.5739  -6.5658
#>   -2.4093 -10.5204  -7.1164   3.3752  -7.3515  -3.9030 -11.0031 -13.9378
#>    1.0782  -4.0710 -20.6551  -2.3617  -2.2396   7.0410  -3.1481  -4.8583
#>   -1.9296   8.6235  -8.1023   9.5233   3.5581  -8.6591   5.4551 -11.1468
#>    0.7985   1.6342   3.0338 -13.6525 -11.6827   1.8645  13.4323 -14.8856
#>   -7.4199   1.4760 -11.3172   9.9795  11.9409  -0.3767   9.0765  17.0295
#>   11.0128  -2.7232   3.7887  -0.7223 -14.9948   8.6560 -12.8197  -4.3247
#>   -8.9076  -4.5126   7.6184   1.6336   3.1292  -2.7199  -8.1854  -1.3475
#>   -2.6407  -8.8116  -7.0053  -7.3184   7.6983  -7.9423   3.6256   5.7701
#>   -4.4484   2.9671  13.4572  12.5593   4.4906   9.5039 -23.6726  -4.8650
#>    3.3939  -4.9087   1.3588  -8.1726  -0.0383  18.4516  -8.8886  17.9514
#>   -6.1902 -11.9312   6.6570  -4.0247  -8.9388   0.2975  14.5608  -1.9238
#>   -5.2678  -9.9522  -4.8622 -12.1945   5.2736   0.3730  -2.6076  15.6229
#>   -1.4450   0.7092  -4.4965   8.4108  20.6527  -9.6488  -7.5758  14.4595
#>    6.2394  -2.4573  10.4891  13.6373   1.9341  15.4998  12.3268  10.6602
#>   -1.3839   1.8595  11.5423 -17.8907   3.4932  19.6371  -9.9388   6.2422
#>   -3.1122   1.2172  -0.2186   7.9200  13.9686   9.9803   4.8530  -1.6604
#>   -6.3423   6.0345  -2.5554   7.7513  -3.2216 -14.3803   6.9449  -0.6258
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]