Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.5913  -1.7270  18.6534  -1.8956   1.0642  -8.4084   9.5673 -12.7987
#>    5.6053  16.4454   6.3376  -6.3205 -16.8858  -0.9430   6.3542 -15.2979
#>   -0.6111  -6.8242  -1.7485  -5.1329 -15.0775  22.6517  -5.0732  -1.0243
#>   -2.4002   4.0610 -10.6319  -8.3054   8.5754   7.9589  11.1653   3.1410
#>   -3.4130   3.3775  -5.8382 -10.6556   2.0752  -1.0424  14.4929   1.7366
#>    6.1563  -1.5442   3.7738   2.8102   0.7144 -10.7556  -9.6810   5.3017
#>    1.7346   4.9395   0.2872  11.0538  11.1826   7.4956 -17.2875   0.7784
#>   -3.6059  -0.7003   2.7671  -1.4555 -11.0490   2.1986  13.4046  -7.5939
#>   -0.1299   4.5358 -15.6872  15.5322  -2.2750 -12.9079  -4.4093  -5.8129
#>    1.7564   4.0345 -11.0241   0.5996   2.3598   5.3553   7.3051  -1.6079
#>   -4.8249   1.3599  -4.2249  -0.8354   7.8308  23.3103   7.1471  -5.7011
#>    0.3030  -1.6680  -0.0713  16.7526  12.0207  -7.6464  -2.1075   4.1568
#>    4.3207  -2.5311   6.2522  -7.7628   2.7106   3.1145  -0.6504  18.8756
#>    3.0715  -2.4335  -0.5573  -3.6440 -15.3692  -6.7380   1.3715  -4.0985
#>   -6.7245   0.7814   6.1247  -8.2303   4.7808   6.6836  -6.1130 -11.3021
#>   -8.4500  -9.6761  -2.1110  -5.5485  13.5269 -20.5682   1.1124 -11.5148
#>   -2.9402  -1.0194   4.3711  13.7220   5.0768   6.1629   8.9586  -1.3227
#>   -3.2702   9.6413  -6.7584  -6.2579 -11.6993  -3.3354   5.9762  -1.4167
#>    0.8545  -1.2488  -3.4370  13.0408  14.4795   1.7013   4.7087   1.3133
#>   -3.2720   4.9596  -3.0884  10.0416   2.7968  -2.5031  -7.1965 -13.9880
#>   -4.8113  -3.3220  -2.1905   2.0324  -6.8335  -7.7756   2.7463  17.7200
#>    0.7835   0.7233   3.6152  13.1998  -1.4759   3.6167  -2.0917  -1.8920
#>   -0.7088   4.5328   3.7995 -13.4933  -6.9102   3.2729  -4.3535   2.4784
#>    1.3385   6.8220  -9.6744 -11.4664   3.6277  17.1675  -6.1533   6.4032
#>    2.0326  -3.7108   1.4216  -7.1579 -12.3947  -5.1028  17.7560  10.3313
#>   -3.2162   1.8463  -4.1332  -7.4710   1.8020  -6.0835   0.8155   1.2041
#>   -5.2538   2.9475  -6.9372  -2.1379   3.7654   6.2474  -4.9191  -2.5846
#>   -0.9762  -0.4092   6.5075   2.3840   0.7176 -11.7121  -9.1626 -10.4657
#>    2.7391  -0.2960   1.8023   8.8584   5.6351   5.9184 -15.7415  -0.3439
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]