Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.7983  10.9981   9.3318   0.8312  -9.3810  -6.2481   2.5174  -5.2121
#>    4.4459   2.5712  -4.4729 -18.1814   7.7537  -8.5841  -5.6727  -1.7714
#>   -4.6735   6.5739   8.1946  -2.7949  -9.2977  -0.2433   1.3373  -1.0378
#>   -0.7885   3.0063  13.7261   4.1016   3.8790  -5.6931  -8.3097  10.1307
#>   -4.5779  -4.1173  -7.9815   2.2084   2.1423 -12.1658  -3.4560 -13.3567
#>   -3.3576   7.4056  -3.0767 -13.2867   9.2849 -10.0478  -5.0380   8.4018
#>    0.5210  -4.9244   3.5613   6.8164 -22.6168  -1.6500  -8.5114  13.2389
#>   13.2928  -5.2754   4.4584  -5.8159   6.9818  -1.4103  -2.5442  -2.5912
#>   -8.2044   0.4961   9.4133  -5.3121   0.7371  -5.5322   7.1463   6.0581
#>   -5.4184   3.6122  -0.0639  -3.2395  -0.7268   1.0669 -10.4738  -3.1754
#>    0.4097  -4.9169  -1.5503  -8.4841  -5.1958   5.7599 -10.8382 -17.8774
#>    7.4320   9.6713  -3.0566  -5.0848   8.3107   6.3893   7.4622   4.7478
#>    0.4772   1.0108  -8.5820  -7.0148   7.9555   3.4893   6.2651   4.6317
#>    3.9464   3.5100 -10.0776   2.2844   9.3877   0.7042  -2.9128   1.2685
#>   -7.0470  -9.8565  -2.9126  -0.7849  -4.5046  10.4036  -3.2090  -3.8884
#>   -0.5147   1.1739   7.3157   7.1520  -4.3947  -2.6434 -10.3894   6.3104
#>   -6.3814   2.1845   1.7497  10.5428   6.4576   2.2330  -5.1480   1.6042
#>   -1.9076  -6.5175   1.4915   8.5897  -0.3831  -3.8062   2.8133   3.0299
#>   -5.4823 -10.3946  -1.2355  -4.4695  -1.8166  11.2657 -11.7399   6.2843
#>   -0.9577  -0.6102  -0.0163   4.9437  -7.1174  -6.3805   2.9400   8.7677
#>   -0.6665  -2.6265   3.2410  -3.2227   5.6186  -2.3369  10.6700  -0.1558
#>   -2.4167   0.7896   0.4426   5.3503  -5.5593   6.3101   4.3545  12.2037
#>   -4.3063  -6.3043  -0.7884  -7.5120  -2.8430  10.7030  -8.8285  -5.7634
#>   -2.0152   0.4544   2.1608  -0.0042   9.0395   2.1649   7.2239   5.1632
#>    1.3916  -2.1503  -9.6091   2.0033  18.3922  -0.0403  -0.9482   0.8779
#>    4.0285  -5.5629  -0.4176   9.8068   0.9941   0.6764 -13.8459   8.2168
#>    0.3243  -4.6714  12.9296   3.6498  -5.3644   2.1509   3.8612  -4.9149
#>    1.4190  -4.8111  -4.7548   9.0414   4.0936  -3.1435   1.4145  -4.0500
#>    2.8077  -0.9672  -3.1046   5.8435   1.9744  -4.7020   5.0696   1.1618
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]