Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   2.0353  -0.0556  -1.0474  10.3000 -18.0804  -0.5622  20.3601  -2.7098
#>   -0.9752  -1.3512   9.8238  -7.6383   2.3378   7.6187  -9.8296  10.9988
#>    4.2060   2.5969 -10.6180   3.2716   1.7670   3.0500  10.8469   2.5095
#>   -3.0240  -2.1891  -0.6180  17.2172   9.1679   0.5183   2.4152  -0.7359
#>   -8.5260  -0.5884  12.6832 -10.1241 -12.7523  -2.5912  -9.5661  10.6227
#>   -1.6996  11.7464   5.6183 -13.0076  -3.0308  -4.2253   5.5737   6.6203
#>   -2.0782   7.0481  -6.7250  -8.8051  -2.2189  12.4422  -0.9122   7.8246
#>    0.1703  -1.3661 -13.8240 -10.2041   6.9426   3.3911  -5.3442   8.0322
#>    0.1074  -7.4990   0.6845 -12.7797  -1.0999  14.6171   7.0003  -5.0298
#>    2.5137  -0.4457  -4.6006  -4.0297   0.1286 -12.6748   6.2805  -0.9486
#>   -3.6880  10.0888   1.3363  -1.9236  13.3648  -3.8794   7.9953  13.2993
#>    3.5087   1.3432   7.5831  -3.3913  -1.7542  -9.9676   1.8834 -17.4650
#>    2.9477  -1.0685   6.8773  -4.3815  -0.4785   5.0936   5.1744   0.6368
#>   -3.5970   7.9771   1.0770  11.3176   4.4574  13.8969  -1.0561  -7.3064
#>    2.4261  -4.3524   6.0132  -4.0944   8.4843   6.3609  -3.3814  -5.2243
#>   -3.2283   3.9594   4.6595  -6.9321   0.9111 -18.7788 -12.7102   2.0657
#>    0.4440  -1.6776  -6.7026   4.9954  -4.0855   3.9521   9.1249  -7.7399
#>   -1.6023  -2.3830   4.6110  -8.7383  -1.3701  -3.0509 -15.3228   2.8361
#>   -3.3989   0.1031  -2.0814  -2.5758 -11.3813  13.8809  -2.2582  -3.4164
#>    0.0070   4.6664  -3.5154  -6.7735  -6.4504 -13.5839  11.2655   6.0363
#>    2.8252   3.1165  -4.7579   1.8836  -3.4393 -13.8305  -1.4585   4.2342
#>    1.8585   0.9265  -4.6865   3.7104  -4.2317  -4.1075  -0.1818  -1.3909
#>   -5.0767   3.8791   0.1328   4.4305  -0.5428  -4.0919 -20.1992 -10.8342
#>    2.1235  -6.8077  -6.5092  -3.6992  15.6660   3.2493   9.8672  22.4307
#>   -0.3230  -2.6205   3.1938   0.8514   3.4083   6.3286  -6.2625 -15.4473
#>    1.6712   2.8402  -1.5507  -0.8933  -2.7344   3.7997   8.0418 -20.8169
#>    0.5710   0.6064   1.4103  -3.6643 -15.7163  -2.3862   3.4752  -2.7726
#>   -0.0151  -1.7025  -1.7289  -2.6764 -13.0278  15.6308   2.9457  -4.9303
#>    5.3141   4.7290  -6.2215   3.8676 -11.3431 -15.1787  -3.3351  -2.2367
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]