Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  8.8842e-01  1.1859e+00 -7.1259e+00  5.2011e+00  5.9045e+00 -6.0688e+00
#>  -2.5283e+00  4.5914e-01  6.9673e+00 -1.6061e+00 -3.6947e+00  7.9439e+00
#>   5.2880e+00  1.8138e+00 -3.8433e+00  4.3562e-01  8.8203e+00 -1.0959e+01
#>  -2.6956e+00 -1.5732e+00  1.3329e+00  6.6290e+00 -7.0131e+00  6.8345e+00
#>   1.9920e+00  1.3906e+00  5.3510e+00 -5.0408e-01  1.0368e+01  2.3485e+00
#>  -1.3899e+00  8.3894e-01 -1.1380e+00 -3.4058e+00  2.5567e+00 -4.6016e+00
#>   6.4483e+00 -1.9851e+00  1.7499e-01  9.1351e+00  3.9293e+00 -2.4419e+00
#>  -1.0644e+01  4.5101e+00  1.2311e+01 -5.2963e-01  1.8590e+00  8.2484e+00
#>  -3.5727e+00  4.6283e-01 -7.0054e+00  1.1428e+00  3.1652e+00 -6.0200e+00
#>  -2.4948e+00 -2.7112e+00  3.1680e+00 -5.7285e+00 -2.8731e+00  7.9332e+00
#>   4.5705e+00 -9.4696e+00 -2.9230e+00  8.1349e+00 -1.2156e+01  3.5881e+00
#>  -1.3054e+00  5.1816e+00 -3.9416e+00 -8.4336e-01 -3.4016e+00  4.4439e+00
#>   3.2948e+00 -1.0852e+01 -4.8006e+00  3.5456e+00  7.9348e+00 -4.7322e+00
#>   6.0845e-01  5.6312e+00 -1.9629e+00  4.9517e-01 -3.2689e+00 -2.7389e+00
#>  -3.3106e+00  1.5897e+00 -2.6407e+00  1.2291e+01  1.1172e+00  3.2616e+00
#>   1.2566e+00  1.9097e-01  4.5645e+00 -4.1141e+00  8.9182e+00 -6.6613e+00
#>  -3.2203e+00 -3.4164e+00 -7.6474e+00  3.8905e+00 -2.1594e+00 -5.9722e-01
#>   8.4378e-01  6.3467e+00  4.8416e-01  8.0593e+00  6.5228e+00 -2.8357e+00
#>  -5.9555e+00  3.3446e+00  1.0473e+01 -4.9112e+00 -3.8369e+00  1.1600e+01
#>   1.2307e+00 -1.8537e-02  1.4354e-01 -5.7790e-01  4.4213e+00 -1.4331e+01
#>   2.1410e+00 -7.8777e+00  2.5348e+00  4.0401e+00 -1.4043e+01 -1.0980e+01
#>  -2.2296e-01  1.3221e+00 -1.5787e+00  5.7657e+00  4.2069e+00 -7.7408e-01
#>   2.8559e+00 -6.0709e-01 -3.2271e-01  4.9160e-01  1.4782e+01 -3.1899e+00
#>  -6.0651e+00  1.3457e+00  1.3216e+01 -2.4811e+00 -1.4419e+00  9.5488e+00
#>  -5.5174e+00  4.9684e+00  2.8490e+00 -4.8312e-01  4.8720e+00  6.2907e+00
#>  -4.4642e+00 -2.0670e+00  3.9837e+00  2.4223e+00 -6.1090e-01  1.1915e+00
#>  -7.4887e-02  9.4501e-02 -8.3772e+00  9.3404e+00 -3.9543e+00 -1.1572e+00
#>  -9.4147e-01  5.9521e+00  7.3403e+00 -1.3676e+01  5.2645e+00  6.0462e+00
#>   1.0965e+00 -3.7988e+00 -4.6849e+00 -3.3456e-01 -5.7296e+00 -2.9761e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]