Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.8839  -2.0286  -7.9941  -4.4912  -1.3919   3.5793  -7.7126   4.9585
#>    0.2695   4.4498  11.6696  -7.7014   0.8781  -7.4230  -5.3592   0.9378
#>   -2.0019  -0.0077  14.8135  -3.2823  -5.2011   2.4669  -8.3689  10.3143
#>    2.6482   3.6757   0.8894   2.7633   0.6045  -7.6503   3.0541  -4.6246
#>   -2.6920   7.1357   3.6183  -4.9887   9.0144  -8.3850   1.3345  -0.3272
#>    0.5377  -2.8119   0.0472  11.0034  -0.9717  -2.8984  -2.9548  -1.8730
#>    2.0956  -2.0377   0.4356  -7.2055  20.3363  -7.0402 -18.4653  -3.3763
#>    3.4147  -0.6326   2.9983  -0.3772  -5.8634   3.9167  10.8256   1.6476
#>   -1.1646  -1.8197  -8.2184   4.3976  -1.7926   5.7375   4.5805   2.8846
#>    1.1343  -1.2416  -9.4278   3.7020  -2.6960  -0.4248  -7.1315   6.4793
#>   -1.7495  -6.8959 -20.2692  -3.0786 -14.8665  12.7289 -11.2771 -15.7268
#>   -1.5105   2.1101  -2.4113  -9.1957 -12.9085   5.5308   9.1700  -0.8114
#>    0.2225  -1.0607  -1.9754   3.6950  -8.9595  -9.2085  -5.8395 -18.2756
#>   -0.1702  11.4856  -0.9017  -1.9841   8.5786  -8.9548   2.0405  -6.2321
#>    1.0658  -6.2118   0.1962  11.2276   3.7513  11.7985  -7.0495  14.4268
#>   -1.6104  -5.1476  -0.3368   0.4314   2.9646  -5.5826   3.6719   4.1772
#>   -1.4546   2.3031   4.5925  -6.9785   2.5787   1.3441   5.0612 -11.2913
#>    2.3167   2.4496   4.1608  -4.4602   0.3731   8.9457  12.0729   1.3457
#>    2.3966  -1.8417  -3.2377   1.4879   4.1955   9.6145  -7.4241  -8.8389
#>    0.2290   0.4615  10.3721  -5.7179  -0.6720   5.8702  -3.4779  10.7572
#>   -2.4632  -3.3842  -3.5956 -10.3454  -8.1997   2.5830   9.6775  -7.1109
#>   -3.4283   2.3099   5.9725   0.2332   3.3272  -2.4730 -14.3884  18.5598
#>   -3.3774   7.0502  -3.5486  -0.8133   3.3382  -9.6173  -3.3829  13.0312
#>    2.5166   3.9356   0.2615  -9.4852  19.8202  -9.1790  -7.2140  -3.2795
#>    1.5202  -4.6688  -2.8197   9.7016 -10.6520 -14.4970  -2.8216   2.6601
#>    2.4158   3.2432  -5.7261   7.2108   2.5393  -6.0320  -4.1962 -13.7893
#>    0.3102   2.9121  -1.6173   4.3049  -0.1345  -2.3454  -3.6099  -5.9419
#>   -0.8185   7.1493  -6.0250 -10.2788  -6.5760  -9.7527  -6.1866   5.7044
#>    2.7142  10.8332   1.5422  11.1784  12.7236  -1.1891  -9.4722  -2.4060
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]