Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -6.6408e+00 -5.1293e+00 -3.1544e+00 -2.6639e+00 -8.6011e-01 -1.3143e+00
#>  -8.7395e-01 -3.2388e+00  4.8989e+00  7.2991e+00  1.7933e+01  3.4038e+00
#>  -7.2649e+00 -4.6882e+00  4.5588e-01  1.0674e+01  9.2522e+00 -1.6784e+01
#>   3.4381e+00  7.9646e+00 -9.6366e-01 -2.6756e+00  4.2752e+00 -5.0821e-01
#>  -1.1317e+00  5.8564e-01  9.1876e+00 -9.7975e+00 -8.2918e-01 -4.6830e+00
#>   1.0943e+00 -1.2264e+00 -1.0366e+01  8.1406e-01 -1.1696e+00 -7.5430e+00
#>  -5.2222e+00  9.4210e+00 -3.3120e+00 -8.3662e+00 -1.2757e+01  3.2359e+00
#>   4.5268e+00  2.7984e+00 -2.9096e+00 -3.4344e+00  1.5183e+01  1.2416e+01
#>   4.7061e+00  9.6342e+00 -6.3280e+00  8.4606e-01 -6.9265e+00 -1.1755e+00
#>  -1.4327e+00 -4.4863e-01  6.9645e-01 -4.2736e+00  8.8124e+00 -6.3562e+00
#>  -3.7767e-01  2.6922e+00 -3.6279e+00  2.0855e+00  6.1891e+00  1.6104e+00
#>   7.8768e+00  2.3898e+00 -1.3127e+01 -4.2543e+00 -5.1739e+00  8.2161e+00
#>  -1.6115e+00  2.1100e+00 -3.5624e+00 -3.2900e+00 -4.7611e+00  4.8626e+00
#>  -1.6820e+00  1.0745e+01  7.4231e+00  1.3669e+01 -3.9154e+00 -3.4704e-01
#>   2.0071e+00  7.0794e+00 -1.4963e+00 -9.0839e+00  3.1433e+00 -2.4823e+01
#>  -1.2713e+00  4.6714e+00 -9.5905e+00  1.0434e+01 -9.1092e+00 -6.5852e+00
#>  -4.0413e+00  1.1356e+01  5.2643e+00  9.9159e-01  3.9639e+00 -5.5902e+00
#>  -1.6191e+00 -1.5557e-01  6.9227e+00  8.5658e+00 -3.3209e+00  1.3852e+01
#>   2.9451e+00 -7.2571e+00  7.0370e+00  2.0094e+00  9.1067e+00 -5.6048e+00
#>  -4.2097e+00  4.7965e+00 -4.9393e+00  5.9320e+00 -2.9090e+00  2.1240e+00
#>  -6.7302e+00 -6.2992e-01 -1.2131e+01 -6.3578e+00 -3.9899e+00  1.8967e+00
#>  -5.1643e+00 -1.1749e+01 -4.5686e+00 -6.9435e+00 -2.4821e-01  4.1045e+00
#>  -8.9654e-01 -8.4651e-01  9.4196e+00  3.3670e+00  1.6574e+01  8.5743e-01
#>   2.4433e+00  5.7832e-01 -7.6168e+00 -9.0036e+00 -4.5930e+00  4.3760e+00
#>   3.0362e+00  3.4515e+00 -1.7832e+00 -4.1572e+00  3.7490e+00  7.3244e+00
#>  -4.1372e+00  4.6742e+00  3.4795e+00  7.8357e+00 -3.7485e+00  7.0962e-01
#>   4.2780e-01  6.1684e+00 -8.1702e-01 -7.6048e+00 -2.4355e+00 -1.3313e+00
#>   6.0796e+00 -5.7544e+00 -3.1076e+00  1.3433e+00 -1.3116e+01  6.2580e+00
#>   9.4917e+00  5.1098e+00  2.1503e+00  1.3062e+01 -4.0734e+00  1.6231e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]