Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.7683   5.1852   3.4343  12.0301  16.1913  -6.2426   5.3045  -0.5273
#>    2.6481  -8.2533   1.3311  -7.7521   3.1230  -0.2887  -5.7633 -10.6329
#>   -6.3727   4.5805   2.9610   1.0962  12.6566   8.8814   4.1381  -6.6540
#>   -1.9473  -3.7825  -5.2646  -2.8579 -10.2717  -5.1233   1.3345   6.2181
#>   -0.9940   3.7841  13.4183 -13.2648   2.0431   1.7543   3.8262  -7.0955
#>   -0.9774  -7.0680   1.0318  13.4609   1.8872 -12.4440  -3.2324 -10.8180
#>    2.8330   2.1420  -8.4816  -7.8551   8.7513  -1.7804  -1.4707  -5.6238
#>   -1.7984  -6.1473   0.6359  -3.9582   8.8878  -0.6184   0.3366   3.1985
#>    7.8966  -0.2853   2.8502 -14.4626 -30.5826 -14.8344  16.5480  -9.9164
#>   -2.7981   7.1795  17.5805  22.7094   4.8040   0.5672  -3.5482  -7.3827
#>   -0.8824  -9.5807  10.2775  -7.0685   9.8833   9.1917   2.7270   5.6922
#>    3.4369  -9.6469  -0.8704  -3.7595   2.8594 -10.4473   7.3694  15.2344
#>   -0.1782  -2.9014   2.1103   2.9653  -1.2927  -6.5416   3.4222   0.1346
#>   -0.8507   1.0517  12.0318 -11.8079   3.7300  -0.3534   5.0051  20.9042
#>    2.2524  -8.5664  20.7770  -2.4199  -0.3150  -4.4218   2.7015  -5.4972
#>    8.3301   1.0525  -2.9155  -4.2826  -9.3125  -5.5405   9.4186  -6.8183
#>    4.4964  -3.2403   5.4496   2.4729  -7.8695   0.3406   1.7341  -3.8617
#>    2.7576  -0.9460  -2.4449   2.9652  10.8923  -1.3637   5.1971  -4.2048
#>    0.5847   8.5103  -1.0952   5.9068   5.4424   8.1342  -2.6735  11.3943
#>   -4.6558   6.6167  -2.6404   1.2418  -5.5092  -1.7549   6.0309  -2.3739
#>    3.1869  -3.5482   3.6358   0.8778  -3.4517  -4.0957  -4.2718  -6.8776
#>    7.8561  -0.5640   5.2865  -9.1322   2.9495   7.7025   2.2610  -6.9696
#>    1.9708   2.5626  -3.3923   4.5870 -15.8462  -6.2581  -1.0501  -9.0720
#>   -5.7600  -6.3418  10.8953   0.4934   6.8904  17.5690   8.6638   0.5337
#>   -0.9897   7.5724  11.8520  -1.7114  10.0698  -4.8284  -8.6576  -9.8788
#>   -2.3414  -9.6260  -3.9092   4.1616   5.9085  14.8476  11.3579   9.3146
#>    9.9989 -14.8676   8.8170 -12.4723 -17.3128  11.8507  -3.9443 -17.0396
#>    1.0623   0.2370   9.2065   2.0593  -1.0056   4.6475   4.3673  -9.3356
#>   -0.3558  -2.4325  -6.9918   9.8670   1.8788  22.1714   2.0045  21.5954
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]