Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  5.8994e-01 -5.1918e+00  4.3052e+00 -1.0964e+00 -4.6168e+00  2.6461e+00
#>   1.1362e+00  5.1513e+00  4.9429e+00  4.8686e+00  2.7967e+00 -1.2545e-01
#>   7.4027e-01 -3.8145e+00 -2.2314e+00 -6.2868e+00  9.5509e+00 -2.0186e+00
#>   1.8193e+00  2.6384e-01  9.2030e+00 -7.3097e-01 -4.5278e+00 -1.5795e+00
#>   1.1188e+00 -1.1496e+00  1.6318e+00  2.2153e+00  3.3637e+00 -1.0906e+01
#>   2.3376e+00 -2.8419e+00 -5.0130e+00 -3.6638e+00 -1.6795e+00  1.4207e+01
#>   3.4157e+00 -6.6530e+00 -2.2546e+00 -3.1423e+00  9.4415e+00  3.6295e+00
#>  -1.0368e+00 -3.0915e+00 -1.1482e+01  1.4965e+00  1.5740e+01 -5.5445e+00
#>  -1.0993e+00 -1.0019e+01  1.2722e+00 -2.3398e+00  1.2132e+01  1.6356e+01
#>  -3.9897e+00  9.5347e+00  5.0502e-01  1.0013e+01  1.9959e+01 -7.3984e-01
#>  -1.1584e+00  9.7720e+00 -3.4870e+00 -1.7041e+00 -1.0884e+00  1.4122e+01
#>   1.7393e+00 -2.1143e+00 -8.8393e+00  2.5004e+00 -9.0812e+00 -6.5361e+00
#>  -2.5311e+00  4.6936e+00 -9.7816e+00  1.1548e+01  1.1255e+01  6.8389e+00
#>  -1.7379e+00  5.1451e+00 -1.0174e+01 -1.7900e+01  5.4573e+00  7.1224e+00
#>  -2.3377e+00 -2.4158e+00  8.5915e+00 -9.3853e+00 -1.5948e+00  1.6941e+00
#>   3.6386e+00 -7.8882e+00 -2.7288e+00 -7.2882e+00 -9.8514e+00 -1.5639e+00
#>   9.3326e-01 -3.7819e+00  1.9493e+00 -1.2444e+00  1.7675e+00 -2.3215e+01
#>  -1.5273e+00  2.4384e-01  7.1467e+00 -8.8541e+00 -5.4116e+00 -1.1231e+01
#>  -7.8811e-01 -3.7107e-02  3.9480e+00 -6.8619e+00  3.7672e+00 -4.4900e+00
#>   5.8097e-01 -9.5097e-01 -6.6887e+00  6.1596e+00  8.3510e+00  1.9997e+00
#>   1.5830e+00 -2.2771e+00  6.2943e-05 -7.5622e+00  9.6848e+00  2.3653e+00
#>   4.1125e-01  2.9906e+00 -1.3429e+01  7.3514e+00 -2.2264e+00 -4.2748e+00
#>  -7.1499e-01  2.7136e+00 -6.3136e+00  6.0063e+00  6.8173e+00  4.8784e+00
#>   7.2385e-01  2.3206e+00  4.3748e+00  2.5004e-01  5.9094e+00  1.7238e+01
#>  -1.7936e+00 -2.6088e+00 -2.4353e-01 -6.9015e+00  9.6377e+00  7.7743e-01
#>  -4.1423e-01  5.7695e+00 -6.2428e+00 -9.3246e+00 -1.0774e+01  9.6990e-01
#>   2.8136e+00 -1.3883e+00  3.7766e+00 -5.9514e+00  3.4741e+00 -3.6344e+00
#>  -4.0867e-01  2.0898e+00  1.3682e+01  2.3190e+00 -5.9447e+00  2.4663e+01
#>  -3.1222e-01  5.2024e+00 -8.1905e-01  2.3240e+00 -1.9461e+01  2.0990e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]