Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.3295   1.1751 -10.7735   6.2758   2.2202   0.2460  -7.9209 -16.9995
#>   -2.4249  -3.0895   2.6031   0.8858   8.0343 -10.1411 -12.2419  -3.8803
#>    0.4578   2.2655   4.0321  -1.1510   3.7484 -12.5633  -6.3306   4.8395
#>   -2.3523  -0.3109   0.3401   0.2441  -5.8943 -12.5267  21.0693   3.1801
#>   -1.7860   6.5962  -3.4074  19.4548  -7.9905  -2.4937   8.1792  22.8489
#>    2.2480  -5.1949   3.1256  -0.1273   4.1239  -2.1680  15.9330  -9.5002
#>   -5.2324   3.7853   0.6568   2.9600  -9.6102  16.4748  -2.7892 -10.3057
#>   -2.7405   1.4188  -4.2076   3.2298  -1.5432   5.8163   7.1649  -9.7849
#>    5.4226   5.3726  -3.6414  -2.5856  -1.6491  -8.6661  15.3731  -5.7682
#>    1.9394   2.4557  -4.3724  15.4358   7.8507  12.3995  11.5753 -13.9390
#>   -3.3783  -8.1661   2.8055   8.9259  -0.0060  -6.8467  11.2333 -10.3151
#>    0.9819 -10.9203   7.0177 -10.3255  17.1981  -1.1520  -6.9294 -10.5295
#>   -0.3339   3.5789  -0.2273  -3.9543  -1.4149  -5.3953  -7.0823   1.6095
#>    1.6928   3.5132  -0.9078   5.7505 -22.4227 -10.4568  -0.4175  -4.0141
#>    0.3350  -9.2222  -4.2696  -0.1022  -1.3647   9.8524  -9.0631  -5.7513
#>   -2.5990   1.2005   2.4412   3.1242 -17.7470   8.3420  -5.6593 -11.7435
#>   -4.2907   0.0969   5.2874  12.4340  24.4122  -4.5047   4.6722  13.6566
#>   -8.4425  -6.4206   5.2676  13.7927   5.5017   4.8146  -0.9466   5.6121
#>   -2.2153   4.3523  -1.9941  -2.3352   7.1129  12.3784  -1.4056  -3.7285
#>   -5.3853  -4.2434  -9.3653   2.5202 -13.9850   4.7658  -9.6450  -1.0306
#>   -6.2597 -13.5538   8.5046   0.3724  -0.4368 -10.2662  -2.8284   9.9701
#>   -3.7081  -2.8620  15.7625  17.9514 -11.7528  12.7922  -1.8189  -8.6924
#>    3.6608  -1.9132   0.1598  12.5992   7.1022   2.1494 -18.1380  11.9528
#>    3.4660   6.2396  15.3127  20.3892  -1.2351  -5.1022  -3.1918  -7.2837
#>   -4.4679 -10.3725   2.0332 -12.3839  12.1938   6.6471  -0.2459  -3.6934
#>   -6.8333   1.0836   5.1668  -0.3022  -3.5010 -10.8402  12.8204  11.7714
#>   -0.6035   6.9949   4.3648  -6.0979   2.1746 -13.1727   1.9618  -3.4417
#>   -0.5906  -2.8647  19.3476   0.7014  24.0480 -16.3556  14.2342 -10.5359
#>    1.5170  -0.2606  -5.6351  -1.9287  -0.2544  -1.6098   8.9525   0.8394
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]