Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.9230  -3.3363   9.1002   5.4503   3.5690   8.4419 -13.5348 -11.7328
#>   -3.5309  -3.6392   7.0904   1.2762   1.4630   4.2831   3.0955  -5.2369
#>   -1.4786  -2.6574   0.9696   5.5044   5.7777 -14.2829   0.4820  -4.8763
#>    5.7289  12.7160   6.5839  -4.4795  -8.2263 -15.6393   6.5381  20.1102
#>   -4.3453   8.8324  -8.6820   3.6700  -0.9563 -17.5692  10.1960  -6.1756
#>   -0.0242   1.8599   4.9217   8.9811 -18.1821  12.3946  10.0483 -10.1594
#>   -1.7092  -7.9536   9.9550   5.3441   3.0928   1.8435  -2.7181  -2.0381
#>   -1.5696   1.5379   0.2212   3.1270  12.4051  -2.3015 -11.8095  -1.5669
#>   -2.8348   2.6012  -0.6278  -1.3107  -2.9430   7.8628   2.0303   2.7786
#>    6.1723  11.8431   3.7309  -2.7203   5.3956   7.8615  15.5960  -0.5898
#>   -0.6197   3.9554   2.0513  -5.5283  -5.4143  13.0888  11.4493  -3.0406
#>    2.5999   0.8681   5.7866  -0.1882   0.9213   6.3387  -7.4750   0.6560
#>    0.4855 -12.6117  -6.1473   7.1696   4.4668  -4.4617  -6.5854  -3.5963
#>    5.8137   6.5388  12.1118   2.8458 -21.4505   6.3317 -15.3019  -0.0876
#>   -3.5157  13.4442  -8.0361   7.7717  -3.8262  -8.9991  -3.6710   5.2895
#>   -5.7468   0.6348  -0.1585   5.1316 -11.6923   5.3604  -4.9421   7.7502
#>    8.5098  -1.2134   5.0562   1.8675   2.7677  -8.3465   0.8915   9.6854
#>   -1.8928  -2.5238  -0.0890  -4.3409   1.0252 -15.7355   4.2665  -8.6460
#>    2.8721   0.2368  -2.1510  -7.6875  -5.4363   0.8539  10.4050   6.5642
#>   -3.3935   5.6691   2.9514  10.4366  -3.6404   3.5498  -6.0198   5.8249
#>    1.7197   0.1317 -10.1388  -3.5621   3.1398  22.1328  -7.1898  -1.6862
#>    6.7241   6.8415 -10.4780  12.8983  -4.7710  11.1558   4.9635   5.1834
#>    2.7848  -0.5021  -3.1250   5.5738 -16.7623  -3.4011   6.9233   7.4274
#>    5.3663   8.2557  -2.5508   5.7807  -1.2490   7.2103   2.2536  -0.0909
#>    2.6032   2.5309  -3.5103 -15.0449 -11.6985   6.6010   0.6311  11.1530
#>    1.8453  -0.9997  -2.4780  -1.5683  -4.0580  -1.8572  24.4701   9.2397
#>    7.7288  -1.8272   8.0169   1.7538 -20.0071  12.0660  -4.7574  15.7323
#>    0.3912   7.7827  -1.2404   6.7807 -11.8011  13.8781  -9.7362  13.9788
#>   -1.2762   8.6386   2.3114  -4.3476   8.7466  -0.8589  15.8099  -3.0691
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]