Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  6.3395e-01  4.9368e+00  1.8553e+00  1.3789e-01 -1.3238e+00  7.9368e+00
#>  -1.8007e+00  7.7440e-01 -2.8809e+00 -2.9199e+00  3.2275e+00  2.5346e+00
#>   6.8222e-01  2.1757e+00 -3.5473e+00 -1.1711e+01 -1.1351e+01 -7.9482e+00
#>   1.4941e+00  4.8824e+00 -4.7864e+00 -9.5165e-01  4.4023e+00  8.4572e+00
#>  -3.7735e-01  1.0174e+01 -6.8727e+00  7.3040e+00  1.4239e+01  9.0715e+00
#>   3.9705e+00  3.5994e+00 -3.9946e+00  5.5361e+00  4.2095e+00  1.9134e+01
#>   1.0489e+00  4.7933e+00 -1.7252e-01 -1.4542e+01  1.2573e+01  5.3233e+00
#>  -1.1277e+00  3.7694e+00  5.6478e-02  1.8197e+01  3.0097e+00  7.6968e+00
#>   9.5283e+00 -3.2081e+00  5.1061e+00 -9.4088e-01 -4.5778e+00 -9.5477e+00
#>  -2.1662e+00  3.6587e+00 -1.1627e+00 -5.4488e-01 -1.9296e+00  3.4383e+00
#>   2.6095e+00 -6.6340e+00  1.6911e+00 -1.6295e+01  5.3767e+00  7.8835e+00
#>  -7.3055e+00 -2.3531e+00 -1.9899e+00  1.3166e+01  1.2534e+01 -4.6387e-01
#>   2.1272e+00  8.4238e-02 -5.5628e+00 -4.1884e+00  9.0887e+00 -2.4694e+00
#>   1.0633e-01 -1.4320e+00 -8.2173e+00  1.0842e+01 -1.7471e-01  2.3511e+00
#>   5.4931e+00  3.7174e+00 -2.5284e+00  6.3903e+00 -6.9219e+00  1.1975e+01
#>  -8.2463e+00  8.7112e+00  3.8870e+00  2.0608e+00  1.0606e+01  5.3488e+00
#>   2.2251e-01 -1.5961e+00 -1.5036e+00  3.1430e+00 -1.1371e+01 -3.6798e-01
#>   4.7768e+00 -2.6818e+00  1.9850e+00  1.3103e+01  3.9865e+00  6.7002e+00
#>  -3.0989e+00 -9.6325e-01 -8.5289e-01 -6.4364e+00 -2.0441e+00  4.0307e+00
#>  -5.0122e-01  7.6025e+00 -7.7951e+00  1.3801e+00  2.9028e+00 -3.1981e+00
#>   5.4004e-01 -5.7697e-04 -5.7414e+00  3.8530e+00 -2.1528e+00 -3.8810e+00
#>   2.7496e+00  2.3169e+00  4.4824e-02 -3.9777e+00 -4.8596e+00 -9.7107e+00
#>  -3.1832e+00  5.2700e+00 -3.1593e+00  5.6106e+00  1.3678e+00  7.2391e+00
#>  -9.7755e-01  3.4344e+00 -4.4904e+00 -8.0793e+00 -1.2543e+00 -5.6982e-04
#>  -1.1240e+00  6.5310e-01 -4.8144e+00 -2.9535e+00 -1.0371e+01  2.4926e+00
#>  -5.5648e-01  6.9086e+00  3.9747e+00 -2.1245e+00  1.0343e+00  9.2755e+00
#>  -3.2406e+00 -2.4548e+00  3.2424e+00 -6.3661e+00 -1.5991e+00 -2.4170e+00
#>   1.6870e+00  3.8132e+00  3.7690e-01  3.9515e+00 -9.4098e+00  1.2307e+00
#>  -4.7007e+00  1.9389e-01 -5.7887e-01 -2.4291e+00 -1.4649e+00 -6.7332e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]