Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -0.8102   4.0300  -1.2569  -3.3700  -6.8477 -10.4768   1.4522   2.1153
#>   -2.3423   3.6926  14.8086   5.0602  -4.2870  17.8889  -9.3667  -0.6487
#>   10.1358  -0.3836   3.0547  12.6304 -17.0753  12.2675  -2.4497   2.7253
#>    1.8393   1.2197  12.9785  -0.2672  -5.0491   1.2187  -0.8759  19.0901
#>    5.2906  -6.5569  -4.8435  -5.0792   4.6065  11.7939   5.4841   1.6939
#>    3.4531   6.3827   3.8827  -2.5266  -5.8008  -9.9625  -1.1272   0.1994
#>   -3.5867   3.4290 -10.6863  -4.2778   2.7904  -9.3951  11.5946   6.1750
#>   -3.3155   8.7123   2.9505  14.8714  -1.8144   1.9680  12.1211  -0.6336
#>   -2.8221  -6.3169  -4.9911   9.6531   8.0494  16.5706   5.0044   1.8680
#>   -1.6545  -3.0286   8.6335 -16.9105   4.4575  -9.4307   1.8749   4.7094
#>    2.4881   1.4050  -3.7248   1.2352  -1.6822   3.9294  -7.1266  -1.8817
#>   -1.0647  -0.2408  -7.1670  -2.8821  -0.3560   1.4711   9.5765  -2.2669
#>    4.6797  -0.0028   0.3852  -0.4461  -0.3954   2.4579  -8.3024   4.5759
#>    5.8385  -5.7100  -8.2006   1.2027 -10.2623  -3.8292  11.3776 -17.9584
#>   -0.2126 -13.5951 -15.6629   2.9612  -0.1180 -16.0254   4.2440   2.1968
#>    1.8221   0.3851   0.1521   5.2887 -12.0545  14.1729   1.0602   5.2402
#>    8.6688   1.3458   5.8203  10.0204  11.0434   7.7335   8.1699 -12.9464
#>    0.2485  -5.1145   6.5311   6.2823 -20.7004  -5.8879 -10.8769  -8.2077
#>    0.5895  -2.3511   4.2966 -11.9143   1.6172  -7.1783   7.3105   2.8024
#>   -1.2860  -1.2819  -6.0913 -22.1089 -10.4280  -2.2799   7.2224   0.1011
#>   -5.3188 -13.1202 -10.4320  -9.3136  -3.3865  -8.0133  -3.7083   1.8196
#>   -2.0101   0.2363   1.3999   0.7761  18.6560  14.8136  -4.5553   0.5091
#>    0.4393   6.7897  -2.6420   6.3643   8.9915 -14.9677  14.2124   0.5583
#>    3.4579   1.4357   3.9814   0.2160  26.2388  -4.4088  20.7593  -6.5559
#>    6.3936   5.2291  -6.8321   6.3083   1.6784   2.5964   4.1883   2.2195
#>   -4.1505  -3.3950   2.7432  -8.5110  -4.9364   3.3937   5.5218   6.9597
#>   -1.3554  11.7569   3.9958   3.1157  -6.9037   7.7593  -1.9722  -3.2734
#>    3.8074  -1.2017   7.0182   8.5270   3.5020  -0.2595  11.1054 -10.3111
#>   -5.7638   2.3693  -3.5352 -18.5536  18.0982  -3.6414  16.9364   4.5342
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]