Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.1845  -2.0433   8.2020  -7.0945  -8.5091  -9.8453   8.0155   4.6715
#>   -4.2267  -0.9053   5.4301   2.7072   7.2958   4.1807  18.4382  -0.7548
#>    1.1460  -2.2694   4.7559   2.3966   1.3225  -8.8889   5.9748   2.8702
#>   -4.4708  -2.4126  -1.3031  17.1587  10.4685   6.3862  -7.4363 -12.5934
#>   -3.5721  -7.4459 -10.7041 -10.1722   3.3484   6.3182   2.4509  11.6945
#>   -3.3271  -0.6483  -3.8213   1.2490  -8.3151  11.3983 -11.8228   1.6192
#>    1.3766  -2.9872   5.6438   3.2774   1.2025  13.7029   1.5048   0.6498
#>    1.3983  -0.7513   3.7926  -6.4357  -7.6228   3.0453   7.8911   0.4616
#>   -1.5732   2.2788   3.9854  -6.3209   6.5704   2.0947  -7.9026   4.7036
#>   -0.2810   7.0467   4.8551  -6.8965  -9.7256  -3.1746 -14.0324   7.4219
#>    7.1768   4.1146   4.4017  -9.3810  -8.4937  12.9451  -3.4627  11.9787
#>   -0.1520  -6.7099  -7.5129  -7.9445   5.1654  12.3269 -16.6650   0.4866
#>    1.3078  -6.1421  10.1575   6.3370   4.9465  17.0066  -5.5572 -14.3860
#>    2.3189   7.7999  13.0325   6.7221   4.8318   2.8632  13.0833   1.1585
#>   -0.9526  -5.3983   0.1383  -1.7126  -5.8539  -8.3479  11.7887  19.6935
#>    0.6032  -0.2043  -3.1250   3.3273   7.4315  -5.0311   8.0556  -8.0800
#>    0.4638  -6.4784  -0.0889  -4.7476  14.7726  13.6010   7.8519  -7.8792
#>    0.9505   4.2447  -4.2994   3.3012  -2.2641   3.4616  -3.2727  -5.2627
#>   -0.9417  -4.2322  -8.5450 -15.3421  10.1291   1.7234  -1.5127   5.7751
#>   -0.9883  -0.1059  -5.6341 -10.8680 -12.6120  -3.5964 -13.4306  -1.7666
#>   -1.9671   0.8007   8.5345   3.2385   9.7084  14.1195   1.6633   2.8432
#>    1.6698   8.1838  -2.5892   5.5900  -9.3449  -9.5041 -11.0896   3.6596
#>   -3.8027   5.2533   1.4022   3.7536   9.2403   9.4989   3.5894  10.2378
#>    5.4161   2.5920  -2.1395  -1.5788 -15.0522   5.6206   1.1063  -5.0286
#>    2.3790   9.3874  -6.2362  -0.9206   4.9526 -15.7360  -5.2181  -3.4804
#>   -2.2914  10.2026   6.7222  -1.2749   4.2996 -14.3380 -10.7241  -3.9600
#>    0.5120  -5.2903  -1.2790  20.2853  10.0300  12.3196  -6.3042   4.3746
#>    1.6156   3.4211  -5.1305  -8.0504  -9.4552  -1.2572   7.7786 -22.2858
#>   -6.9233   0.6309  -8.9145  10.7401   4.6481  -5.1340   6.4816  -1.9489
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]