Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.6874   5.5055   1.8235  13.5980   2.4160   6.2557  -6.9839   2.1025
#>   13.3259   5.8326 -12.4622  -6.6999  13.6990 -11.2314 -15.5238   1.9531
#>    1.2322   1.8917  -1.4375  -6.7434   2.7779  -0.2723  -1.2107  -4.1424
#>    7.8935   7.7072  -2.1302  -1.9226 -17.4773  -4.6161  -8.5262 -17.7855
#>    3.5212  -5.3182   9.7569  -3.8040  -7.8116  -4.1003  -4.6980  11.0607
#>    1.5039   0.6181  -2.1827  -4.8395 -10.1845  -8.0087  -9.5431  -1.8232
#>    2.8706  -6.4943   4.2196  -9.7930   0.6247  -2.6482   5.0599  -5.9998
#>    2.6933  -3.7137  -6.9924  12.0103   3.8987 -15.2800  -9.3164  -6.0611
#>    4.1615  -1.9257  -9.9570   1.5848   2.4647   1.9294   7.7670   8.5985
#>    5.2188  -2.6720  -2.6834   6.7585  -2.3837  -5.2109   5.1067   4.1891
#>    0.2164   3.2986  10.9736   3.3410   5.2477  -0.4706   4.0818   5.1925
#>   -3.1088   0.3052   8.6268   4.7184  -6.0357  -0.8333   2.7451  11.7578
#>   -0.7596  -3.7596   2.0976  -4.7001   0.9630   6.4323  -4.0270   9.2954
#>   -7.5278 -12.0125 -17.9928   2.7371   2.5898 -13.6442  13.1508   3.2271
#>    2.9186  -2.9274  -0.4911 -12.9868  10.4734   6.8875   5.4067  -1.5530
#>   -4.2684   7.7528   3.8826   1.3321  -7.9594  -1.6937  10.2674  -9.1347
#>    0.6640  10.4618   9.7900  -0.7427   7.2513  -8.7488   9.4598  -7.7947
#>   -2.8635  -8.1233  -1.7966   8.9757 -10.6065   1.5846   9.8408   4.4112
#>   -2.7500  -5.7248  -3.4436   4.8094   1.6981   2.9249   3.0113  -0.9237
#>   -5.3387   3.0464  -8.0598 -10.2219  -0.9506   8.9133   7.8456 -10.4807
#>    1.3762   2.3001   2.0247   5.3643   7.3687  -8.9963  -1.2216   0.8938
#>   -4.6812   3.3368  -0.1723   1.3874  -3.0794 -10.3882  -3.2318   7.6924
#>    0.1763   0.2884   0.8270  10.7764   6.9131   4.5162  -2.5029  -4.0156
#>   -1.0090  -2.7146  -0.8359 -10.4462 -15.1492   9.7861   6.5707  -0.6404
#>   -0.7309  -6.7009   0.8041  -0.4104  -6.7780  -8.7469  -0.2874   8.0326
#>   -0.6150  -7.8538  -5.9938   6.1720  -3.7171 -11.2944  -3.3214   7.5502
#>   -6.0416   7.9698   1.3997   1.9412   6.8963   4.6150  11.2979   9.9195
#>    3.5024  -2.7225   0.7726   9.6131  -0.1876  -6.5508  -8.3899   5.9690
#>   -0.5233  -1.1041   3.6496   2.5119   2.4360   0.9127  -3.9992   4.4729
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]