Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.3523  -0.7390  -1.9062   1.7777   2.6841   5.8782  -3.5349  10.9140
#>    0.0664   0.3125   4.8529  -3.8595   2.8141 -14.7321   4.0095   4.0265
#>    1.8406 -10.9242   2.7200  -2.6389  -1.7129 -10.5235  -2.0044   2.1604
#>   -6.6412  -1.5211  -5.4174   1.4910 -12.8075 -14.7931   7.0615 -24.9674
#>    1.2227  -2.1876   0.7145  -2.4314  -0.8112   5.4854   2.1194   5.0850
#>   -1.6643  -2.2291  -3.6595  -6.4937  -6.7171   1.5268 -10.2568  13.0094
#>    1.1403   2.2732  -2.8295  14.6001   1.1555  -6.9378   0.1474   7.3926
#>   -4.6704   3.2540  -4.1558   8.9682 -12.7198   7.9558  -3.5825 -18.5459
#>    1.9555   2.8383   3.2230  -9.1048   0.4913  16.3813  -8.4198   3.5314
#>    0.8675   1.5946 -10.5491   0.4947  -1.0262   1.5311  10.8841  11.6456
#>   -0.7483  -0.6805   6.7589  -0.9762   7.1007   4.1502   0.5139  13.6840
#>   -0.9991   1.5108   2.9791   2.0346   3.8544  -6.9386   5.7652  -0.7133
#>    3.8450   0.6047   0.9509   4.0722  -5.8147   7.7029   0.6946   5.5401
#>    5.3912  -7.0946  17.5597 -11.1261   2.9290  -2.8071  -2.4109  14.0822
#>    0.3144   7.5529  -0.9119   9.8261   2.9406  -0.1000   5.6997   5.1620
#>   -2.8059   1.3344  -2.2351   0.7122  -6.0737   3.6986  -0.7224  -5.7832
#>   -0.1595  -4.0340  -0.5037   7.6075   8.3294   2.1862   8.7359  -3.5945
#>    4.4311   0.8644   2.2414   5.6947  -8.2390   3.1136  -1.3010  -0.7278
#>   -2.7589   4.2329   2.7517  -2.9302   5.6779  -3.3971   8.9324   6.3382
#>    8.0002  -2.3090   5.0646  16.3107 -11.2556  15.4981   4.7700  11.2375
#>   -0.0362  -6.4520   6.2647  -2.4874   8.5923  15.6458   4.2929 -20.4879
#>   -1.2349  -1.5330  -6.3701   0.0363  -9.8594  -2.9463  14.9218  -4.5460
#>   -3.1986   7.7084   6.5475   9.8264   4.4819   2.5443   1.4277  25.3239
#>    1.0904   8.6709  -1.4880   2.4495  -3.7761  11.0637  -1.4606 -10.3852
#>   -2.4876   4.1903  -1.2783  -4.2227  -1.9724  -1.8435  -0.4631  -8.1820
#>   -3.1639   3.0860   3.1211   6.3189  -5.7004   0.5553  14.0701  -2.4095
#>    0.9565   6.0923   3.0520   5.9716  -5.6973  -6.7970  -1.4938   9.0561
#>    3.3407  -3.7894   3.4726  12.8144   1.7798  -2.4115   5.6234   0.3162
#>    0.5597   6.1506 -10.0068   1.9471  -2.8185  14.8187  -7.7699 -10.3671
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]