Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.9334   4.2285   6.3151   6.1420  -0.6719  10.1207  -7.4172 -14.1078
#>    4.1803  -6.5570  11.2302  -5.1924   5.4610   6.0058   3.8571  -5.4066
#>    4.8417   3.2380  -3.1754   5.5807   6.1215  -1.6857   6.8273  -0.6007
#>   -4.7257  -4.5432   8.1289   1.5020  -5.3518  -2.2970 -12.2691 -14.2571
#>   -1.3259  -0.0017   6.0839   1.6047  -1.3763   6.0278  -1.2633   7.6988
#>   -3.1270  -2.1285   1.0118  12.4183 -21.1240   3.7471  10.9193  -0.4922
#>   -1.4737  -0.8610  -0.5005  13.6370   3.4742   8.6755 -10.2855   0.4551
#>    3.6575  -5.1275   1.6318  -7.3062 -10.4163   2.5057   2.1763  -9.7561
#>   -3.5446   5.4172  -2.5087  17.2813   1.2524  -6.2890 -10.0721  -2.8412
#>   -1.7866   0.1718  -0.7615   1.9162   3.7583  -3.8310  -9.3352   6.3639
#>   -4.1676   4.9884  -1.3892   7.2323 -12.7008  -1.4081   6.2912  -0.8495
#>   -4.2272   7.0333   4.3934   8.1035  -0.4311   0.8848  -8.8594  15.3570
#>   -4.3808 -14.4552  -1.6764  -1.7427  -2.3661   9.2606   1.4383 -14.3722
#>    1.2219   1.6189   3.6073  -6.4212  -9.9659  -0.6192   4.3280  10.5426
#>    1.7131  -0.6707  -1.0284   3.2974  -4.6617  -1.6461   6.4461   6.4209
#>    5.1721   3.9258  -3.1418   7.9545   2.7392  -1.3222   3.2671  -6.9638
#>    1.6091  -2.4634  14.5030  -3.0086 -10.0490  -1.2590   2.1312  -4.3163
#>   -2.0685  -3.1954   0.8094  13.4356   8.5278  -6.1036  -3.3414   4.1129
#>    2.3670   3.0021   6.2497 -15.6522  -1.9751  -2.2862  -8.0368   6.2746
#>    0.1986  -5.5740  -6.7309  14.6888   4.4874   5.0149   7.8851  14.1115
#>   -3.7243  -4.5669   9.1888   1.3108  -9.4288  -0.4430  -8.2265 -13.4808
#>   -2.2210   9.1885  -3.6583  -7.0818  10.2784   0.5638  -7.7238 -11.2425
#>    0.4118  -1.1662   1.3257   9.4661  -4.9189   2.9669  -1.9917  -1.2629
#>    0.9410  -0.4325   5.6111  11.7551   6.5654 -20.2793   3.7171  11.4932
#>   -0.1573   1.1775  -0.6228   6.9490   0.2473  15.0343  10.7209   3.4922
#>   -2.8872  -5.5879   1.8222  -2.8955  -1.9595  -0.9886   2.3455  -1.6524
#>   -2.2483  -4.3999  -5.3488  -1.6046  12.7070   9.2816  -5.2318  -6.2227
#>    1.9168  -5.8812  -3.2307  -9.2528  -7.9993   5.2633   8.0328  -1.5113
#>    1.5120  -3.5907   0.6574  -5.3077  10.0890  -2.5267 -13.5162  14.4950
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]