Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.4064  -3.4372  -8.3703  -1.1427  13.3191  11.1180  -6.4251   6.1667
#>   -4.2628   4.3698   4.9907  -4.0541  -9.8638  -8.8200 -16.2428   7.4701
#>    1.0109  -0.9899  -0.5104  -0.9929  -0.7534  -1.4615 -11.5385   6.9593
#>    4.6926  -0.9141  11.3154   5.5613  -7.5883 -13.4802  -0.8112  -1.8908
#>   -0.0369   0.7266  -6.4014  11.4857 -12.5011   7.5707  15.6797   2.7928
#>   -1.4651   4.0243  -1.1156   8.0434  -5.4971   1.4237   7.4644  -9.0781
#>   10.2410   2.4253  13.2184   5.4161  -1.7033  -1.8099   3.2024   8.3359
#>    5.1240   4.0413  -3.1414  -1.8201   8.4315   8.0551   1.2841   5.8196
#>   -8.8455  -7.8408  -2.0025   5.5353  -0.9669  21.7618  -1.6095  -2.2519
#>    3.5637   0.9845   9.7456  -1.7095  -2.8584 -15.6328  -1.2854  -5.6569
#>    0.3651   7.9673   4.8201  -0.9391   1.0636 -16.6849  12.9423 -21.9426
#>    4.0193   4.7781  10.0516   6.9981 -16.0269  11.1537  -9.6888  -7.7267
#>   -4.3584  -1.5336   3.7210   5.5470   0.8831   6.4420  -4.0786  -4.1470
#>    4.3218   1.3256   4.6243  -3.9737  -6.1203   1.1040  -2.9997  -4.1612
#>    0.8495   0.8138  -8.2594   5.1477 -15.6214   0.5548   8.8056 -10.2512
#>   -3.3025  -3.9365   1.6568   0.4392   9.4945  -7.3046  16.2194  11.8549
#>   -0.3378   0.4589  -3.0251   9.5325 -16.5958  18.8661 -13.6853   0.6824
#>   -1.9342  -3.5450  -4.4975 -11.3764  12.3026 -10.6015   7.4007   7.6035
#>   -6.4445  -5.0455 -13.0424  -4.9474  -3.3540   1.1148  22.3378 -16.2912
#>    3.6277   8.5895   3.2742  -1.9250  12.3078   0.3161  -5.1620   1.4701
#>   -1.1163  -1.6947   1.7363  -4.7575  -9.5319   2.7879  -0.0470 -13.7841
#>    1.8764   1.7241   2.0406   3.2605  -3.5854   2.2736  -5.8962  -5.8406
#>   -3.7992  -8.8328 -12.3661  -4.4342  -5.0415   5.6142  -5.7706   4.5041
#>   -2.8318   3.2434  -7.8542  -0.1818 -11.8169  -7.3933   2.5903 -21.4768
#>    8.3860  -3.6083 -13.8167   7.9118  -0.4326   9.0708   5.5695   9.9399
#>    0.2460  -1.1499  -1.3727 -11.1054   1.3644  17.4107   1.9942   7.2615
#>    8.7634   5.8339   7.4715   8.3506   3.7569  14.2609   9.4630   4.0248
#>   -0.7465  -9.9091  -0.4145   5.3102  -4.1925  -3.9277  21.9495 -24.6106
#>   -3.7743 -11.1323   5.9164   9.4236   7.7983   9.6112   1.7447   2.2462
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]