Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  2.8608e+00 -9.3709e+00  2.4940e+00  7.9957e+00 -1.6246e+00 -1.1139e+01
#>   3.4775e-01 -5.5472e-01 -3.2422e+00  6.0402e+00  7.0345e+00 -1.6124e+01
#>  -3.6216e-02  3.3317e+00 -6.9149e+00 -7.3389e+00 -2.9639e+00 -4.7287e+00
#>  -1.8607e-01  2.4255e+00  2.6137e+00 -1.3523e+01 -4.3343e+00 -1.3748e+01
#>  -4.0763e+00  6.5314e+00 -5.1611e+00  2.5967e+00 -7.8638e+00 -1.0961e+01
#>   2.6014e+00 -2.2521e+00  2.7684e+00 -3.9129e+00 -8.3832e+00  5.5459e-01
#>   3.4590e+00  3.1513e+00 -1.5126e+00  2.4800e+00  3.2559e+00 -5.2341e-01
#>   1.9729e+00  2.5716e-01 -8.7713e+00 -5.0488e+00  6.5942e+00 -1.7121e+01
#>  -2.3776e+00 -4.3209e-02  4.7847e-01  2.9628e-01 -1.0358e+01 -1.8171e+00
#>  -3.8125e+00  9.3219e-01 -7.8645e+00  1.2552e+00  3.9206e+00  6.8251e+00
#>  -1.8765e+00  2.0756e-01 -6.3120e+00 -4.7043e+00 -2.6571e+00 -2.9569e+00
#>  -1.4646e+01  5.3400e+00  8.4310e+00 -9.5953e+00  5.7301e+00 -1.5970e+00
#>   4.4578e+00 -4.2486e+00  5.9442e-01 -9.3884e-01  1.4611e+01 -2.0591e+00
#>  -4.5519e+00  1.7767e+00  8.1798e+00 -5.8726e-01 -5.2314e+00  6.9799e+00
#>  -5.1098e+00  5.2809e-02 -6.8278e+00 -1.1842e+01  2.4667e+00  1.4277e+00
#>  -3.8360e+00 -7.4731e+00 -6.9185e-02 -6.8776e+00  1.2214e+01 -1.1878e+01
#>   2.2802e-01  1.7847e+00 -6.2682e+00  1.0232e+01  5.4286e+00  4.6442e-01
#>   2.1474e+00  2.2504e-02  4.6533e+00 -9.2275e-01  2.5383e+00 -6.0811e+00
#>  -6.3368e-01 -3.5685e-01  5.0837e+00 -7.3393e+00 -1.1245e+01  1.7458e+01
#>  -4.1779e+00  4.2052e+00 -3.4884e+00  3.7420e+00 -6.0462e+00 -5.7227e+00
#>  -3.8323e+00 -8.6858e-01 -3.1082e-01 -5.2191e+00  6.1864e+00  3.6073e+00
#>   8.1842e-01  5.4358e+00 -1.4294e-01  4.1074e+00 -1.0618e+01  7.5343e+00
#>  -1.1950e+00 -3.0676e+00  7.6517e+00 -3.2633e+00  7.4986e+00  8.0531e+00
#>   2.6446e+00  7.5390e-01  4.6444e+00  1.7436e+01 -4.8165e+00  5.7693e+00
#>   7.3599e+00  8.5664e-01 -5.8871e+00 -1.1151e+00 -6.9338e+00 -1.1030e+00
#>   2.8651e+00 -2.4296e+00  4.6789e+00  4.5176e+00  1.4570e+00 -6.6800e-01
#>  -3.9265e+00  8.6853e+00  4.6159e+00  6.8102e+00 -1.8353e+00 -4.4285e+00
#>   9.1729e-01 -4.6665e+00  3.3887e+00  3.5135e+00 -1.3784e+01  8.0815e+00
#>  -2.1898e-01 -3.0009e-01  5.8349e+00  3.1380e+00 -5.4977e+00  9.9600e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]