Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.8123  -5.8381   6.1395  -3.1764 -11.3480  -0.6901  10.4572  -3.8456
#>    0.6941  -5.3832  -1.2213   4.7370   7.9153  -8.2975   3.9453   2.1852
#>    0.6572  -4.3824   9.5955   1.1875   3.2416  -8.6507 -18.2054   2.9490
#>    1.2533  -1.1357  -1.9960  -1.9853   1.3105  -1.3300  -4.1739   4.7351
#>   -3.1845  -2.5156  -0.5865   8.4148   3.9448  11.6268 -23.8634   2.3488
#>   -5.2826  -0.4585  -6.9971 -10.2506 -11.9157   8.0103  -2.4839  -5.6584
#>   -0.1183   2.2262 -10.0734   5.7629  -5.1995   9.6982  -0.3483  -0.2912
#>   -3.6124  -2.8272  -3.4964  -9.1936 -14.4607 -17.2623 -16.2479   2.7096
#>   -3.1363   7.7704   5.4513  11.9727   8.0430 -16.5252   4.2477  17.6856
#>    1.7091  -2.5777   3.5851   1.8558  -6.6023   8.5403  11.7955 -17.6128
#>    2.3927   2.2877   4.5246  -9.8726 -13.1480   2.6506   5.3788   9.1246
#>   -0.2352   4.6819   5.3545   9.9485  23.0152   4.6019   9.3314  15.5102
#>    2.9109  -7.5840  -3.5760 -10.8928   6.4324  24.9105  -1.0540   1.2120
#>   -3.7348  -3.9791  -0.2651 -10.9514  -2.2202 -12.7320   7.4393   6.2840
#>    2.2665   5.2107  -0.8960  -1.4726  -7.1801  16.2771 -10.7368 -18.5499
#>   -2.7512  -1.6295 -12.0583  -4.4127  -3.3091  13.2887  -8.0685  10.4560
#>    2.4166   1.0656  -3.2869  -3.7329 -17.5316   4.7615   0.7345  -6.2714
#>   -1.1992  -0.1196 -12.0830  -4.7962   1.4793  -8.3027  15.5694  16.9517
#>    0.1394   0.4346   9.2791  -1.9256   3.4248  -3.8487 -17.2751  -1.1546
#>   -1.7014  -3.2948   1.4250   7.5664   3.2521  -3.1570  -8.6085   1.0949
#>    0.4600   1.5417   6.6583   8.2894  10.9009   7.0822  -0.0507   4.6634
#>    0.0323   0.3277  -0.0480  -1.9167  -2.6807   3.2697   6.4867 -12.3970
#>   -1.9838   2.1026  -0.3564  19.4293  18.4832   1.8752   2.9070   4.7640
#>    0.3027   2.5832  -2.7179   7.6593  -2.4518 -16.0687   8.4062 -14.9143
#>   -4.3790  -6.5723  -2.9165   9.3007  -7.3025  12.3313   3.4382   2.5424
#>   -0.8617  -3.0233   5.4875   2.0546  12.5160   5.4350  -3.7597   0.6506
#>   -0.5663  -2.5571  -5.5933   1.2845  27.3700   3.6062 -16.2888   4.4038
#>    3.3005   9.7075   7.8359   0.5955 -15.0268   0.2698  -1.2497  -2.3809
#>   -0.7577   1.0420  -2.7391  -1.8040   2.5930   0.8825   1.1716   9.6408
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]