Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.1692 -12.5665  -7.7316   3.2600 -16.1729   1.3458   3.2534  -9.9382
#>  -16.2769 -12.6311   4.8749  -2.6461 -21.5933   2.8185 -16.2314 -10.0859
#>    0.5911  -7.2148   1.5888   7.1947  -0.4513  -5.7478   1.2479   8.7092
#>   -1.3436   1.3967 -14.1825  -3.9316   2.1171  14.1369  -0.1818 -14.4536
#>    2.8858   0.0287  -6.8441 -10.0593   7.0586  -4.7568  -8.7100  15.0830
#>   -5.0143  -6.7791  -5.5566   4.0748  -7.8663   2.2379   7.7692  -2.1548
#>   -4.9522   2.3237  -7.1158   3.3288 -13.9892  -6.9620 -13.4923 -10.4351
#>   -0.2689   4.9817  -6.6388  -5.3178   2.4373  -2.8619  -3.1219 -12.9966
#>    7.7371  -2.6624  -3.8092  -3.8608  -2.3194   7.0411   2.3633 -22.8854
#>    2.0458  -0.5345   8.4641  14.4054  10.0641   2.4466   7.9527  12.0745
#>   -1.1230  -3.8136  -3.3668   4.1344   6.7287 -10.4098  -1.0387   6.0069
#>    3.2985  -1.3999   9.2925  19.5441  15.4359   6.6789   8.2415  -2.7583
#>    1.4184 -10.8452   4.8378   8.7710   4.8380  -8.3177   0.1486   5.2298
#>   -5.5202  -9.6062   2.6820  -3.4668 -12.9809   7.2385 -13.7684  -5.9977
#>    8.0245   7.7299  14.9070  11.6243  -4.7301   0.7414   2.0971  11.3964
#>    1.8917   7.7141   5.9679  -7.3523 -11.5155   3.8052  -2.6299   5.1632
#>    5.5277  -7.6407   9.9423  -1.3746 -14.1347   7.5702 -14.9625   7.3535
#>    7.5845  -6.5943  -1.1134   1.8991 -14.9860   6.7989   8.6815  -9.0276
#>   -7.1617  15.6968   2.0572   4.8318   3.3646  15.0527  13.4173  11.0036
#>   -1.2064  -1.4610   1.1720 -11.6624  -8.1303  -4.2381   0.4795  11.3497
#>    2.1547  -7.0049  -6.2379   7.0577  11.1374   6.9152  -5.8201   3.4252
#>   10.0628  -1.3907   6.8824   1.0052  -1.1045  -3.8093   8.5600   0.8944
#>   -7.1787   4.0474 -10.2314  -6.9939  -6.8964  -0.5830 -11.0454   0.4355
#>   -6.2469  -0.8075   1.7025   2.7788  -4.0060   5.7187   4.9412  -1.0524
#>    7.4459  11.0525   8.9892   1.3103  12.4586  -5.6804  -2.1625  -6.1622
#>    3.2054  -1.9008  -5.9249 -13.2486 -13.0777  -5.2888  -3.7290   7.3893
#>    0.2600  -5.9121   6.9591  -2.7320 -10.4932   4.0981  -0.6616  -7.5113
#>    0.1180   9.3739   0.2749  12.8246  13.4486  -2.2315   8.8162   0.5205
#>    3.9020   2.9509  -3.6641   3.9175   8.0535  15.4841  -7.4792  -9.2671
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]