Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   5.7848   2.9619  -0.3042  -4.2420   2.5002 -12.4975   3.1185   3.1041
#>    3.3634   1.3227   0.1806   6.5194   1.9321  19.0780   0.4568   5.5163
#>    2.9262  -2.1448   2.9686  -9.9705 -10.0264  -3.2756 -16.3078  10.9311
#>   -2.8190  -7.5357  16.5415  -1.7000  -4.6998   4.1735  -8.6541  -0.6978
#>   -1.4303   4.8914  -7.9845   3.3767 -10.9601   0.2731   8.4613  -6.8245
#>    2.7886  -6.9181   2.6312  -0.3392  -3.3148 -12.9909   8.1302   3.6846
#>    5.5047   6.2879   3.1488  15.9983  -0.8607  -0.4833  -3.9151  -6.0696
#>    5.2181  -3.3011   3.0383  -5.7440   0.7480   7.3770  -7.9037  12.2828
#>   -2.8196  -3.6956  -5.8821  -2.2326   5.6639   2.9640  -2.2119  -6.2074
#>    7.2239  -1.9257   0.9453   5.9865   2.1672   4.4350  -6.9397  -8.2121
#>    3.0989   9.2105   1.2883   7.3219 -14.8291   9.4709   2.5051   4.1970
#>    4.4126   5.5534   7.8202  -4.8046   6.8401  -4.1859  11.4859  -9.0956
#>   -5.4673  -5.2713  12.1501  -5.0947  -4.7211   0.7333 -15.2553  -4.1999
#>   -7.3779  -1.3845  12.6026  -1.5721  13.7759   9.3767   4.4665  -2.3721
#>    2.8786   9.3662   0.8987  -4.5699  12.0625  -0.0587   0.2809  -2.2282
#>   -1.7732  10.3230  -1.0774  16.7010  -1.8254  10.6353  -2.9311   8.6805
#>   -0.6172   4.0993  -0.1372   2.7708  -9.6549  -2.1258   1.3237 -14.2016
#>   -6.3633   3.3870 -11.8756   4.0007  -3.4717  11.8729   9.8322  -3.0302
#>    1.7622   1.2841   3.7486   1.8436   8.4495   6.8017  11.7928  -2.9630
#>   -5.0623   0.7866  -9.8152  -6.1256  -4.1890 -12.8123  -4.7716   3.7112
#>   -0.6389  -1.1001  -5.8113   4.1271  23.0664  -6.4357  15.7793  -0.2706
#>    2.9911   4.7985   7.3241   3.9552  -3.9934  -2.2407   4.1498  -5.6681
#>   -2.4355  -4.4754  -8.4791  -9.4491  -9.2816  -2.4655   1.2887  11.6670
#>   -4.9642  -5.1417   5.5250  -3.1687  -7.3800  -4.9425 -13.9159 -13.0618
#>    2.2445   7.2912   4.0042  -3.4588  -2.5041  -2.7390  11.0109   4.9040
#>    0.0845  10.0697  -0.7287   7.9435 -10.5075 -11.7865   2.6924 -16.6281
#>   -0.5513   3.7784   4.7208  -7.5804  -3.2449  -4.5101   0.2068   7.2136
#>   -1.7051   0.5379  -6.1919   7.1102   6.1290   9.3593   5.2795   4.5024
#>    3.9676  -0.8544   7.6490   1.4312  -5.0166 -10.6913   6.1940 -11.4884
#> ... [the output was truncated (use n=-1 to disable)]
#>   10.4442 -17.2131 -11.1833 -21.4063   4.1492   0.0329