Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   4.6903   3.5188   7.6153   1.1876  -8.6252   8.1962   5.0679  -2.9113
#>   -1.9607  -6.6698  -3.2292   1.0773  -7.3052  -6.1716   4.2019   4.9168
#>   -1.8445   1.3931   3.2532  14.2035  -3.3876   8.6903  -8.0308   2.4673
#>    6.0733   1.0809   8.4201   1.4341  10.0801   7.8919   0.0253   2.1555
#>    0.6225  15.4599  18.3464   1.7858  -6.2995   3.7856   4.2261  15.3930
#>   12.3171  12.9280   1.0300   2.7825  -8.3223   1.4133  -6.1587 -11.3144
#>    9.3573   3.1550   3.4809   0.4895 -14.8227   4.4735  10.8998   0.7497
#>    1.6366   0.1984  -0.0879  -0.2026   6.1272  -3.7022   4.9211 -12.5193
#>    2.6869   8.2579 -16.0062   6.0781  10.2900  -1.8434   8.6953   8.4113
#>    0.3169  -1.9988  16.4096 -13.6150   8.5627   0.4329   2.9720   7.8565
#>   -2.4317  -1.3239  -8.3043  -5.5666  -2.8576  14.2396  10.0877  10.1115
#>    3.8766   2.3761  18.3892   0.3549  14.9997 -14.3325   3.7221  11.2288
#>    0.6026  -1.5735   7.3022  -3.1865   7.0865  12.2740  -4.2654  -0.2551
#>   -6.4232  -3.6214  -7.9500  -6.0634  13.8299   9.5509 -17.9530  -3.3083
#>    2.1038   4.1206 -12.1308  -3.3671   3.0084   3.6488  -2.3004 -14.8938
#>    2.5256   0.5982  -2.5861  -9.9434  -8.7062   4.2377   9.3086   5.7035
#>   -7.2121  -5.3719  -8.2275   0.5887 -15.4429  -1.4873 -10.6088  -1.0757
#>   -5.3622  -0.8109  -7.2337  -8.6268   2.8339  -8.0493  10.3099  16.4707
#>    1.3759 -16.1251   0.6848   9.1371 -10.9534  -2.5569   8.8548  -7.5924
#>   -3.8118 -15.7673   0.7712   0.0605   6.7482  -8.1682  -0.9654  -5.5145
#>   -1.1724  -0.0834  -6.3884   8.4079   5.3435   6.6169  24.0344  -0.5279
#>   -2.8207 -10.1038  -0.3561   0.5174  -8.0413 -10.4949  -7.5754   1.0579
#>    4.5265  -6.3861 -16.1239  12.1753  -6.7559  10.3864  10.6914  -3.3647
#>   -1.7812   4.4872   2.8502  -4.0206   0.1413   1.0244  -9.0909   0.9224
#>   -2.1343  -1.1535   6.1486  -0.2527  13.1858  13.2386  -3.1482  -0.4628
#>    2.0545   3.6662   2.8217  -0.1132  -8.5407  -4.2578   0.2760  -9.0510
#>   -0.5504   3.4621  -8.0101 -11.6344  -3.1073  -7.4768  -7.0619 -11.7850
#>    0.5485  -0.8804   6.1687  -1.5108   6.6341   2.8182   2.4598   0.2027
#>   -9.3169  -4.1349   9.0987  13.3497  -3.0994  15.0614  13.1643 -10.5454
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]