Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.1521  12.9776   9.8473 -12.8403  12.5707  -5.7610  -5.1330   4.0443
#>    0.8875   3.9218   6.0341   6.1875   6.7605   1.4862   5.8527  -0.7782
#>   -0.1432  -0.2358  -6.8665   5.9821  10.3553  -3.5820   1.0262 -13.3974
#>   -3.3816   3.1193   5.2919  -8.0842   1.1295 -13.3039  -7.7515   0.2686
#>    3.5155  -4.6341   9.2506  -6.5191  -0.7902   1.6335  -3.2651  -1.1553
#>    7.2535 -17.0180  -0.3725  -7.5507 -13.5426   9.4424   2.7911  -3.8838
#>   -2.3306   5.9876 -18.1735   1.9628 -19.4688  -1.8620   0.7287   3.4827
#>    1.9483   3.3849 -13.0630   4.7245   1.5794  12.7802  -0.0128   7.9275
#>   -1.4188 -15.1385  13.3919 -12.1504   8.0161  14.1379   7.4942  -3.2942
#>   -7.6999   1.3747 -10.5088  -8.2934   4.1743  19.7915  -6.9036  22.5696
#>   -3.7265   4.7299  -3.4924  -6.0016 -19.5015   5.0688   2.5652 -11.7069
#>    1.0687  -9.3006  -0.3861   9.1267  20.3886   9.4130  11.1560  13.6940
#>   -0.0213  -2.1716   5.1551   9.8374   5.4378  -0.8600 -16.1939  -3.2340
#>    2.0677   9.6286   1.0863   7.4184  -3.3148   9.1268  -0.4667  16.9016
#>   -9.4500  -0.1420   9.6799  -7.4846 -10.8891  -5.2448  -3.5071 -13.8056
#>   -5.5126  -4.0044   9.5970 -10.9397  -4.6203   2.9685  -4.1850  -1.9544
#>    0.1445  -7.6576  14.8652  -2.9988  -3.9774   0.3406  -2.4834   1.6615
#>   -6.5199  -0.2915  -6.3189   6.7414   3.0011  -1.7713  -8.8677   6.6575
#>    4.2990   0.3493 -16.7536   3.3461  -4.8512   6.3271  -3.7979  -7.6482
#>   -4.2587   6.5203  -2.7887   7.3346   8.3010   0.1549  -4.7908   4.8314
#>   -3.7495   0.3768 -13.7430   0.1011   3.0674  -0.1701  -2.7002   9.0690
#>    0.9071   2.3130   2.7541  -6.1432   7.8134 -13.2196   7.4805   1.1532
#>    4.2941   5.5949   1.7396  -9.5869   2.7481 -24.5547   4.1146  16.8972
#>  -11.9267   6.7836   8.9469  -9.6038  20.3106 -28.6632   4.3642  22.8899
#>    7.5494  -2.7628  11.8850  -8.3591 -27.8957  12.7713  -6.1348  22.8087
#>   -1.8128  -1.9709  10.2704  -4.4175  -8.4771  11.4838   6.8504 -18.6557
#>    3.6353   4.3326   2.8781   2.4605   3.1918   0.9499  15.3836 -11.8924
#>   -1.8239   4.9755   6.0363  -1.9617   8.2978   0.4179  -5.8900   5.7312
#>   -3.1517  -4.8311  11.7080   6.1292   1.0817  -2.3725   1.0665   6.3584
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]