Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   1.4355   4.9602   5.6583   5.7073   4.5956   3.1290   2.2239   7.2011
#>   -3.8044  -8.0601  -2.5401  12.9368  -0.3337  -3.8498   8.8412  11.1920
#>    0.4247   2.4279   2.0597   3.9653  14.7747  -1.3618  -1.8162   8.5838
#>    4.1718   5.0583  10.9622   5.1634   2.4457   3.4336  12.2930  -4.9376
#>    2.5724   0.6032  -3.2435   1.5624 -13.1505   1.3677   5.1562  -2.8484
#>   -1.9939  -4.4339 -12.0674  -7.8084 -14.9726 -10.4067   5.1703  -6.3095
#>    0.4936  -3.6718   8.6278  -3.7043  -3.6432   7.5071   3.5436 -18.4593
#>   -2.0982   0.0408  -4.5761  -7.0758  -2.8911 -10.8539  -3.0617   8.1224
#>    5.3656   8.8097   2.1163   2.6138  -6.7366  -2.7833  -6.4776   0.2399
#>    0.5865   1.4767   1.6523  -4.6355  -0.6680   3.5481   6.0012  -6.8879
#>    2.9098  -1.2233   5.1790  -1.1422  -3.3527  -1.4165  -9.1260   8.4739
#>    2.1239  -0.9337   8.9044  10.3523  -3.1288   7.1579   3.6944   4.9349
#>    9.2018   8.2407   9.5928   1.7531   0.7846  -8.7060   5.1628   0.9165
#>    4.5097  -6.0057   5.2184  -6.3763 -10.4064   0.1025 -14.5763  -5.1367
#>   -5.0270  -2.1696  -2.2514  -8.9170  14.9498   1.8575   1.1392  15.9217
#>   -3.4738   2.8193   2.4857  -6.9388  12.3356  11.1089  -1.6874   6.1251
#>   -0.3709  -2.3178   5.5074  12.8701   4.9117   7.1518  -6.3883  12.8017
#>    4.0805  -4.1742  11.2483  -6.0659   5.4581   7.0074  -7.4701 -24.0541
#>    0.7306   0.5980  -2.6529  -3.0331   8.1327  -8.9623  15.8067  -6.7202
#>    1.2045  -9.7853   2.9125  -1.9202  10.3756   9.4293   8.7776  -3.6974
#>    2.3575   2.8359  -6.3603  -0.4023  -9.4884  -3.7510   9.4929  -1.4482
#>   -0.3939  -3.6168  -0.2834   6.0281   4.6860  11.6863   3.0223   4.0122
#>    1.5888   0.6033  -9.6627   3.4929   4.8656  -3.7343   8.0472  22.5658
#>    1.1699   2.1418  -7.2450  -7.0824   1.2284 -16.0341  -7.1815 -12.9660
#>   -2.6536 -10.5758  -3.3231   9.1256  -9.5982  -8.0083  -5.7192   6.4427
#>   -2.5696  -2.0968   0.9918   1.5913  -7.3149  -8.6572   0.5086   0.4365
#>    2.9850  -1.7820   3.8588   7.2612  -2.1789   5.9065   7.6463   2.2245
#>    1.8292   4.2629  -5.2031   1.3850  11.4252  -3.7643   8.5670  -2.1287
#>   -1.2114   3.7369   6.1559  11.6816   7.6274  12.1634   2.9661   9.8925
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]