Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.9149   4.1763  -2.8941  15.9712   0.8640   3.6916  -9.0570  11.1466
#>    0.1794   5.0331   0.6750 -15.6481   7.8314   8.8470  20.9138 -13.7304
#>    3.3710   2.3508  -1.7089   5.8880  -2.3482 -13.6753  12.4607   2.6700
#>    2.7949   4.1126   0.4355   8.6668  -2.6202 -10.7796   5.6736  -7.2577
#>    1.4502  -4.0048  -9.4777  -4.4337  -2.9023   8.4074   6.3166  10.4302
#>   -3.5803   4.3841  13.7103  13.2227  -0.0568  -2.3085   0.9643 -12.9046
#>   -3.2541  -7.0067  11.8338   0.7843 -14.1828  -2.1760   5.7131  14.6713
#>    3.3831  -4.8496   3.0971 -11.3062  -2.1317  -4.3177  17.2623   1.0252
#>    3.3788  -1.9966   1.3208   4.2740  -2.7799 -19.7530   5.1040  12.1398
#>   -0.4391  -0.3477   4.0144  -0.0527  -7.5007  -7.3857   6.7010   6.4542
#>   -2.7165  -5.9389   2.9983  -9.7310  11.8621 -13.6006  22.7592  12.9356
#>    5.8602   3.1142  -6.3574   0.6414  -4.8508  -4.6181  -9.1275   5.1252
#>   -6.2945   1.1385  -3.7079   3.2150  -2.0300  17.1394  -0.6326  -0.0213
#>    6.0158   6.5174   1.1374  -1.5344  13.4779  16.2932  11.5972 -12.9987
#>   -4.8380   2.5475   3.7963  -6.0990   6.3774   0.9901   1.0479   7.2047
#>    3.0102  -1.2993  -9.1374  -6.8363  11.8074 -22.4517  -7.3770  -4.0543
#>    5.0742  -1.3543  -1.9936   3.3884 -11.1919  -9.8382  -6.9433   7.2298
#>   -0.0350   4.5529   2.3841  10.4549  -1.9697  -1.7285  -7.2613   7.6169
#>   -6.5413   2.8529  -2.3566 -15.8653   1.1221 -13.9360  -5.4488   3.1239
#>    2.0412  -3.1582  12.3585   6.9168  -3.8944   4.1787   0.0213   2.7357
#>   -7.4920  -0.6913   8.5230  16.0670   9.7869   0.3654   1.2970   2.1838
#>    0.6943   2.3894   2.6701   1.3173  -5.8169  -1.0071   6.0403  -8.7089
#>    6.0544  -1.2893 -10.6560  -2.1465  -6.8482  13.5902   3.4419  -3.9051
#>   -3.2200   3.2209   0.8085  -7.3858  -3.0157  13.3436  -8.2629   4.0599
#>   -2.4531  -5.2925  14.0257   5.0595  -3.4387  -0.5908  -2.0953  -5.8239
#>   -0.2405   5.5854  23.5590  -8.2490   3.4494 -13.1038   7.8169   6.3572
#>    2.1899  -3.0669   4.4320   6.8391   1.1019  -0.4546  -3.8410   6.9204
#>   -3.8163 -12.0967  -8.7149   0.3987  15.1133  19.5653   4.0598  -8.6726
#>    2.1670   3.9602   0.0223  -4.5592   7.2942  -7.0059   5.4881   5.2894
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]