Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.8985  11.9747  -5.5985  -2.5468   6.9687   1.7833  15.7665  -0.0017
#>    0.6903   2.1839  -6.6845  14.5371  -0.3250   7.7990  -9.3703 -12.0997
#>   -2.9981  -5.1172  -1.4802  -1.7967   4.3187  15.3943   4.2331   5.1263
#>    5.5628  -3.4972   0.2763  -5.5056  15.2060  -5.0236  -2.0410   8.5581
#>   -5.8878  -0.4438   1.0628  -1.2906  11.2478  -4.7079   6.8560   0.6744
#>   -2.6403   8.1292   5.2214 -10.6506 -11.1420  -4.0268  -2.4284  -5.6436
#>   -4.1922   0.0224   2.8000   2.1899   0.2900  -9.2219   1.6140  -4.5994
#>    4.7852  -5.7838   2.3860   4.2711  -5.0734  -4.2027   3.3426   4.8954
#>   -6.3825   5.7435  -9.8056  12.4725  -5.6826   1.5911  -9.1934  16.4519
#>    3.2508   4.0989  -2.7135  -6.7411  -0.3535 -10.2911   5.4759  -7.8637
#>    0.6504  -0.6184 -12.2162   0.0919  -5.3382   4.1992   1.4446  10.5975
#>    1.0972  -6.7294   4.4674  11.4670   8.0642  -2.2542  -3.1656  -1.3462
#>    2.7946  -5.0236  -3.9676  -4.4803  -1.8865  13.9416  -5.2259   6.9379
#>    3.5934  -7.8372   5.5732 -17.8260  16.1401   1.9225  -2.9839  -5.6002
#>    4.9677  -6.8604   2.9884  -4.1794  -0.4060  -2.2022   4.8040  -6.2966
#>    4.0046   3.4045  -0.3806   9.6148  -7.9171  10.9609   0.6361   8.7429
#>    0.4474   5.0182  -4.2213  17.6441 -12.8014   3.6292 -10.7072  -0.3461
#>   -0.7143   4.7049  -1.3773   5.3616   5.9267  -4.6100 -10.6349  14.4869
#>    2.6543  -4.0689   3.4926 -12.3366   2.0929   2.8229  -2.1997   3.9475
#>    3.6334   0.0834   4.1723  -7.0122 -10.6590   0.7863   4.9320  -1.9750
#>    6.0872  -2.8235   3.2025  -6.2413  -0.8239   4.1163   5.0376  -0.6820
#>    0.5840   7.6721  -7.9399   0.5218   5.8003 -10.4061  -4.4886  -0.4163
#>    1.6031   0.3755  -1.6904   1.9716   7.8086   9.6469  -4.3104  11.8753
#>    3.2901  -0.2062  -8.2752   3.9940   7.1409   2.8096  -4.6575  20.9578
#>    1.7189   4.6977 -10.8663   0.1773  -4.8017  -7.8723  -1.5289  -8.0703
#>   -3.5524   1.1725   0.7087  -2.0182   1.0671  -9.9547  -3.8758  -2.9470
#>    3.8348  -6.4027  -3.7803  -7.6111 -11.4919   6.6358  -3.5639  15.3433
#>   -0.6937   6.0772 -10.5241  -4.2876   6.6337   3.4290 -15.2187  -4.3099
#>    0.1095  -3.0166  -4.8953  14.6014  -1.3028  -3.8994   7.8260  16.1820
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]