Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.0880  -4.9673  -2.7852   1.9759   5.0683   6.0605  -3.5724   6.1032
#>   -2.2292   0.8292  -4.6512   1.5462   2.2246   3.5337   2.1435  -1.2657
#>   -0.6646   7.1607   2.5247   0.8747  12.6314   1.0797   0.5074  -7.0639
#>   -2.9887  -1.9554   4.8787   9.6449   6.3596   5.4053  -1.7249 -16.5930
#>    2.0883  -2.7523 -13.6431  -8.0725  -2.1588   3.5368  -6.0763  -5.4317
#>   -1.6564  -8.2850  -4.8327   8.7393 -14.3676   9.8998  -9.6820   2.4155
#>    1.2162   9.2923  11.8464   1.1487   3.2445   0.1369 -11.8699 -12.4482
#>   14.8937  -0.2627  -2.6997   5.7789  -2.6390   1.7200 -20.6201  -1.4790
#>    6.6466  -2.5089   5.5280  -6.9862 -23.9882   6.2115  -9.3124 -17.6895
#>   -0.7173   6.8490  -0.6885  -0.1507  -0.8540  -5.4720  -9.3052 -11.6742
#>    1.8652   9.4325   9.4468  10.7921   9.2892  -1.0564  -0.5123 -15.6117
#>   -3.2929  -5.2015   4.3797   0.3543  -1.5105  -3.5879   7.8149   7.0986
#>    4.9765   2.0275  -1.2053   3.7669  20.4404   7.5723   2.8941  -7.2901
#>    2.6804   6.5887   6.6212   1.1579  22.1821  -6.2297  -3.4837  -1.2197
#>    2.5876  -2.5286   6.0301   3.7747   3.4991  22.9623   4.5998   5.5213
#>    3.2784   6.2478  -3.8061   1.7217 -12.6427  -2.3355  -5.6700  -9.9864
#>   -4.4677  -0.6239   0.2254   1.6390   8.0013  25.3737 -15.0163  -0.5345
#>   -4.2319  -1.6379   1.0402 -15.2503  15.3814   0.0473  -7.1110   1.0664
#>    0.5932  -0.4226   7.3127   3.9771  -9.4101   6.0393 -10.9554  -3.8040
#>    1.6555   4.3796  10.3877  -0.4375 -10.5174  15.0691  -3.3660  -2.0422
#>    3.7168  11.7397  -0.9473   0.9298  -1.6471  -2.4521  -6.5935   3.1715
#>    5.2025   1.4794  -2.3299   5.9189 -10.6940  -3.7100  -8.8899  -2.9083
#>    4.8289  -3.3472  -6.3842  17.9562  -4.1500  -0.6232 -12.4665   5.5750
#>    4.7546   1.3217  16.6362  -4.2153  17.3771  -1.6702   1.3246  -8.8231
#>    6.9682   4.0761  13.6124   8.9470  -6.1179  -4.5467 -18.2749  -8.3140
#>    0.9513  -2.7676   3.7642  -6.4950   0.7815 -12.3044   7.3451  -0.9947
#>    5.5584  -3.2340   2.0481 -11.4384 -17.9091  -3.8190  -1.0147   8.8043
#>   -2.4485  -5.0172  -0.6658  -1.4254   5.3903   0.8675   5.3309  -5.0039
#>    7.2239  -9.7703  -4.6378  -3.5398   9.9750   4.0097   3.4043 -15.8685
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]