Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  8.0015e-01 -4.1963e+00  3.8693e-01 -1.3031e+01  6.9693e+00  4.4382e+00
#>  -1.0276e+00 -8.6245e+00 -7.1336e+00 -1.2103e+00  9.3574e+00 -2.2152e-01
#>  -1.8348e+00 -9.8518e+00 -2.8894e+00 -7.1728e+00 -1.0880e+01 -1.8957e+00
#>  -1.5821e+00 -1.9118e-01  2.2589e+00 -6.4085e+00  1.0400e+01 -5.6517e+00
#>  -8.1408e+00  5.5168e-01  8.8325e+00  1.0771e+01 -8.9738e-01 -2.4473e+00
#>   5.9414e+00  4.6249e+00  7.7852e-01 -1.3519e+01 -9.3810e+00  7.4827e+00
#>   7.1199e-01  8.0626e+00  6.6035e+00 -2.8156e+00 -3.9507e+00 -1.3247e+01
#>  -2.8020e+00  3.8686e+00  7.2380e+00 -8.7012e+00 -8.3194e+00 -1.2801e+01
#>  -4.0698e+00 -5.2714e-01  3.7796e+00  5.5228e+00 -9.1128e-01 -2.4073e-01
#>   3.8541e+00  7.1105e-01 -1.5983e+00 -8.8859e+00  7.4699e+00 -3.8406e+00
#>   1.6159e+00 -2.8185e+00  1.9340e+00 -4.9325e+00  9.4727e+00 -3.7164e+00
#>   1.8685e+00  8.7940e+00  6.2698e+00 -4.6071e-01 -7.6092e-01 -4.1667e+00
#>   5.4082e+00  5.1307e+00  1.0403e+00 -1.4651e+00  3.1510e+00  1.0699e+01
#>  -2.1899e+00  2.8543e+00 -8.3483e+00 -6.3718e+00 -1.0855e+00  6.7606e+00
#>   1.3442e+00 -2.2659e+00 -6.2866e+00 -5.8099e+00 -2.9253e+00 -2.6635e+00
#>   3.6224e-01  7.0982e+00  5.9395e+00 -2.3688e+00 -1.1984e+01 -1.8534e+00
#>   2.6274e+00  2.2892e+00  6.6934e+00  3.5290e+00 -6.0985e+00 -9.2460e+00
#>   1.0829e+00  1.6698e-01 -1.3120e+00 -3.0315e+00  4.6781e+00 -6.3532e+00
#>   9.0461e+00  6.7670e+00 -1.7200e+00  2.4939e+00  1.1473e+01  7.0480e+00
#>  -7.5099e-01  2.8571e+00  1.2262e+01  5.6587e+00  7.0981e+00  6.8610e+00
#>  -2.6542e+00  9.1341e+00  9.9624e+00  3.9994e+00  7.7920e+00 -1.8427e+00
#>   1.0360e+00  6.4262e-01  1.3455e+01  1.1814e+01  1.3605e+01  6.7462e+00
#>   4.4222e-01  3.8817e+00 -1.7632e+00 -1.4316e+01 -9.5142e-01 -2.6729e-01
#>   1.8189e+00 -5.4563e+00  8.2271e+00  6.1725e+00 -2.5930e+01 -6.6448e+00
#>   3.4838e+00 -3.5689e+00 -8.5240e+00 -8.5999e+00 -8.4005e+00 -2.8494e+00
#>  -6.4920e+00 -1.3430e+00  4.8061e+00  4.6412e+00  8.2770e-01  5.0632e+00
#>  -5.8370e+00 -4.8244e+00 -8.0498e+00  1.2573e+01  5.9466e+00 -1.4521e+00
#>   2.2230e-01  6.3075e+00  1.0009e+01 -7.5698e+00  2.9992e+00 -1.3310e+01
#>   4.0752e+00  2.7310e+00 -5.5631e+00 -1.0509e-01  1.8596e+00  5.7408e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]