Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.0588   2.1429   8.7332  11.0517  21.9029   1.3037 -13.5168  13.9520
#>   -0.4828   7.4469   2.6998   2.7694  -4.1023  16.7502  17.8524  -4.7451
#>    0.4988   1.5344  12.1825  -0.9782   5.2676  -1.4582  12.5997  -3.6015
#>   -0.3582  -5.1260   4.2277  24.4835   1.8371  -7.4173  11.0409   2.5570
#>   -2.5543  -4.6734  -1.3585   4.4925  -0.0805   5.6596  -7.5445   2.7206
#>    1.5470  -4.4740   2.6591   1.8266  -9.7858  -9.0867  -8.5055 -10.8729
#>    1.4245  -0.8213   0.0888  -2.1392   7.0977   8.5046   3.3862  -3.6375
#>   -4.2504  -1.2546  -2.3592   7.6158 -11.8369  16.4763  12.4933  -6.7137
#>   -6.7805  -4.1029 -10.1605  -3.1287  -2.4827   2.4839  -9.1788   8.2102
#>   -0.6040  -2.4742  -0.5388   2.9655   2.0418   3.3542  10.6715 -13.3938
#>   -2.3585   6.2006   3.3977   3.2846   4.1561  15.5809  12.7289  24.5208
#>   -1.1240  -2.4225   2.7673   2.4384  -3.6435   2.5117  -3.7041   0.9565
#>   -2.5789   1.9589  -7.0554  -3.9496  -6.9015  13.9396  -7.4156   4.2497
#>   -0.7946  -3.8742  -3.8692  -5.1239   0.6836  12.1362  -2.3528   0.0154
#>   -2.7240   4.8721  -6.0857   5.7191  -4.7695  -1.2963  -9.9754   4.6829
#>   -2.3366  -9.2231 -16.2960 -10.3296   0.2488   6.6668  -6.9965  -4.1230
#>    0.4922  -1.5286   4.7055   1.9008   5.1502   1.3317  -2.9653  -1.6850
#>    0.8268  -5.1132  -2.3871  -0.6781  -4.0358 -10.8554 -11.1395  -5.8168
#>    4.4076  -8.0703   2.9437  -9.5581  -2.7385  -2.1244  -5.4011  -5.6084
#>   -1.2437  -2.3942  -1.1184 -13.3541   0.9957  -8.6106  -7.4791   5.4621
#>    6.9159   0.3732   6.7677  -3.9790 -12.6934   6.2991   5.4742  -7.0369
#>   -1.4103  -0.7070  11.3935  -2.0238   1.9280   6.0297   0.2843  -2.0675
#>    1.8002   1.5316   4.0748  -4.3520  -4.9121 -11.9655 -13.6873  -2.7128
#>   -1.9311   7.1899   7.0404   8.2224   5.8477  -8.3176 -15.5254   9.9366
#>    3.3902   2.4166  -3.6214   4.3582  -2.9424  -6.7208   2.9653   9.4672
#>    6.3207   5.9362  -0.2420   0.4705  -5.2983   9.2653   1.1315   4.8732
#>   -2.3524   7.6039   2.2340  11.5724  -7.3289 -10.3364  -9.6481   8.3344
#>    0.0933  13.3185   2.4282  -1.7095  -6.3846  -4.4142 -17.7057   0.2569
#>    3.1862  -6.0075  -9.7540  -9.8189  -4.5000   1.4940   6.7194  -2.4906
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]