Conv_transpose1d

Conv_transpose1d

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}

#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -3.2614   4.2061 -15.4228  19.5765  -2.2262  14.0118  -4.8554  -2.7092
#>   -2.2540   4.2053  -0.8771  -5.1081   2.0219  -2.6928   0.4380 -12.5340
#>   -4.5138   8.4777 -17.9380   9.1120  -4.1171  -5.7266  -5.1140  21.6639
#>    1.8343   3.8327   7.8695 -17.9234   8.7380  10.4573  -7.8161  -8.0238
#>    1.6575  -5.3940  12.3098  -8.9931  13.9953 -13.9258  19.1461   0.3436
#>   -1.4029  14.3253  -4.5019   9.3621  -8.0106   7.0606  -0.1445   6.9809
#>    0.3005   0.3081   0.8763   5.2392  10.9458   0.0475   5.5526  -6.5026
#>   -0.1077   2.4213  -1.5551  12.5917  -8.9935   2.3630 -11.7416   5.6720
#>    3.0140   4.8140   1.3012  10.4828  -3.5183   2.7476  -5.6964   6.1490
#>   -0.8321   7.7094   2.1945   7.8350   4.9203 -25.5747  -9.4570  -9.8647
#>   -2.7675  -1.7409   4.8086 -24.4729  11.3273 -14.8455  -1.0399  -7.9267
#>    1.0966   1.2623   8.5856   5.2905  -3.1608  -4.9458  -9.4598 -12.2135
#>   -6.0393   2.0766  -1.0955  11.0454 -24.5435 -10.1356   7.3637 -13.4367
#>   -2.9524  16.5498  -6.4839  15.0224 -27.2387   4.3426   4.0403  -1.7358
#>    1.8766  -0.8032  11.1071  -8.8493   9.4369 -27.5891  -4.2580  -8.4046
#>   -2.9927   1.4483  -4.8944   5.6196   0.9391  -2.9300   5.5588  -9.6584
#>   -5.6358   0.1769  -2.2970  13.0686 -10.2487  -5.3287 -22.9864   3.7172
#>    1.4049  -8.2534   2.8561  -1.8350  14.0712 -13.0230  -5.0114  19.8064
#>   -0.7453   8.0611  -4.1917   7.6203 -26.3135 -13.7311 -15.1924   2.0888
#>   -1.0524  -0.4331   9.2080  -0.3699  -2.5487  -3.2863  -9.4579   3.7152
#>    3.9047  -8.6178   3.2256 -10.6290   6.3542  -1.7713  -8.2703  -8.0839
#>   -1.0338  -4.9111  -3.0639   8.1236   3.6839 -12.8180  -9.7243  -9.3385
#>   -1.0135  -2.2951  -5.5044  -5.2055  11.9858  -6.5746 -16.5589 -14.6507
#>    2.9025  -5.9905  -7.3503  -3.5912   9.2372   6.1928   0.6166  -2.1845
#>   -0.6974  -1.9186   0.8521  -7.2266  -4.3058  -7.5573   1.2868 -13.6587
#>   -1.0262  -3.5245  -5.3509  -9.2523   1.9731  12.0488  15.7129   5.0426
#>   -0.2538   2.0648  -6.4379  12.6440  -2.0444  -0.1199   7.2899  -8.3577
#>   -1.5485   3.4367  -1.6910  13.6745  -2.4846  12.5135  11.2871   8.2472
#>    0.7643   8.2355   0.9858  10.2482   9.9030 -17.0416  -0.7318   9.4204
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]