Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -4.7221   9.4097   3.4383   0.4334   0.4731  -9.6288  -9.7699  -0.7864
#>   -5.2781   1.9965  -7.1601  12.6670   7.5510   3.7335  17.2963   2.3405
#>    2.1328   5.7401   3.8702   1.8246  -5.8343   0.2713 -10.7322  -7.1988
#>    0.8131   0.8989   1.8699   8.4550  -5.3010  -7.7439  -5.1573  -7.9531
#>   -1.5442  -0.7710   5.6205  11.3472  -4.9638   0.2704 -13.8275  -6.9121
#>   -3.4877   3.2740  -9.0563   7.1648  -8.5094   6.9864 -18.2323 -10.2756
#>    0.7392   7.3924  -7.1786   5.9072   9.2437   8.0755   2.1173  13.7140
#>    3.8388   7.9059  11.9711  11.8198  -6.1818   6.5727   7.2644  12.6986
#>   -2.3767  -1.3253   3.6046  -4.0939  -9.4625   6.8899  -0.5941  -4.0343
#>    0.7293   4.0426  -5.5795  -0.9297   0.5717  -0.2554 -12.9906  -6.9631
#>    2.2715   3.4637   6.7193   2.9789  -1.8056   7.4486  -0.5307   3.3388
#>    3.4276   7.5635   4.7851   0.8910  -8.0068  -4.1286  -7.5648   4.3404
#>    6.4243  -5.3565  -0.4667  -0.5418   6.4107   9.4854  11.0856   7.5954
#>   -2.3286 -12.2032  -3.9329   1.3880   5.0654  15.1458  -4.8384   3.9749
#>    1.6744   5.9380  -1.0927  -5.3312  12.1597  -0.2663  20.7324  -3.9358
#>    4.2286   7.9342   2.5917   1.7986 -12.9327   7.4555  -6.7219   9.4425
#>   -2.0420   8.5889 -11.4534   4.3770  -5.9435  -4.9813  -3.8264  -8.6446
#>   -1.2110  -4.1079   4.9529  -3.1064   6.8282   0.1106  -1.5953  -5.9675
#>   -0.1167   3.1039   1.6985  -1.7339   8.2943  -1.1904   1.5277   1.8691
#>   -1.9517   5.6977  -3.7057  -7.0795  -3.0107 -19.7487  -2.7348 -17.7029
#>    4.1534   1.5111  -0.6810   5.6995   4.2550  -1.9910  -2.0347   0.3513
#>    2.0540   2.8862  -8.7286   4.6232  -4.5158   0.3317   5.5112  -7.0699
#>    0.7287   3.2077  -2.3889  -5.5246  -7.9011 -15.0795  -1.9878  -8.1093
#>   -1.6994  -1.7751  -0.7589  -6.2679  -1.9795  -2.6687  -5.2372   7.5952
#>   -2.9807  -6.1000  -0.9187   6.4213   4.4677  -0.4981  -0.7806 -15.2533
#>    0.8662  -0.8443   2.4110   2.0660   4.8157   0.3206  -5.8967  -4.8018
#>   -2.5893  -0.9855   3.2017   3.7917  -8.3104  -1.0867  12.2659  -1.3352
#>    4.5456   1.9482   5.1547 -10.9495   5.3042   3.3122  10.7339  15.2285
#>    8.3507   5.4298  -3.8916  13.3810   0.8017  -5.1126   6.8959   8.9048
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]