Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.9242   5.0837 -16.8389   8.3265  -5.5803  -0.2772  10.0249  12.4787
#>   -4.7782  -4.4094   8.3282  -1.1520  12.5795  -1.6304 -10.2569  -2.0657
#>   -1.5000   7.6587 -13.4132 -17.8507   2.2995  10.7768 -20.1129   5.0396
#>    5.3765  11.6229   5.3989  -0.9004  -1.1765   1.7196   8.6137   3.7779
#>  -10.1649   9.4423  -8.3185  -1.1343  13.7728   4.2859  -4.1211   1.2596
#>   -1.8281   0.1450   3.1163   1.7913  -7.8530  -2.7619   1.4105   8.8408
#>   -2.1578   8.3784  -0.4573   5.8762   1.4551   8.2516  -2.4966   1.0074
#>   -0.8119  -2.1622  -3.2335   2.0501   1.4117   5.8780  -9.1133  -4.2604
#>    5.2180   0.2849  11.0126  -1.9341  -2.3397  18.8973  -6.6051   5.1330
#>   -1.6338   6.1479  -9.7118   1.9578  -1.3214  -1.3147 -11.1346  13.7740
#>   -0.3078  -5.9344   0.3271   5.0426   5.2036  -1.1202   1.1446   2.6484
#>    1.7101  -0.0555   8.4006   0.5670  -1.2061   2.1363 -11.0541  -4.0702
#>    2.3652  -5.9965  -9.2162  -3.1100  -9.6810  -3.1707  10.8535  14.0337
#>    2.0942   1.4247  -6.9007  -3.8484   3.7879   1.7696 -10.9637  -2.3984
#>   -4.0106   6.2305 -13.7478  12.8872   9.3579  -4.1019  -5.0794  -0.3848
#>   -0.7188   7.4651   2.4269   1.0226   8.3929   0.5363  16.6326  -7.7068
#>   -2.3495  -4.3169  -6.9905  -2.6972  -5.7397   5.0318   1.6415   6.5151
#>    1.1331  -0.1095  -4.2676   5.4044  -0.1392  -6.7757   7.3027  -3.6177
#>   -5.7639  -1.9252   4.8617  -4.6279  -8.6052  16.9188   7.7716 -10.5991
#>   -0.5884  -4.2829   9.3469  -1.1823 -13.1125   0.9592  -1.4959  11.6160
#>   -4.0091  -4.3115  -3.9035   2.6370  -9.6365  -5.1399  -2.6696  -5.9998
#>   -5.6283 -11.3722   2.8140  20.6990  -4.9654   4.2315   6.0019   0.5063
#>   -2.4446  -1.2231  -6.4218  -2.6279  -6.5127   2.2063   8.6236   2.2362
#>    1.2752  -4.3126   8.5313  -2.2282   4.3138 -15.8682  -3.0879   6.1621
#>    1.0701   2.0364   8.2267   0.5607   7.8258  -5.8441   6.6814   3.2527
#>    5.4859  -5.9378   7.3787   6.5228  -9.7773  -3.1064  10.6240  -4.4966
#>   -4.0965   4.8358  14.4100   2.5899   2.2445   8.8831  -2.1125  13.2298
#>   -0.2794  -6.3877   3.2522  -7.5161  -6.2306 -12.2597  14.8828   5.4223
#>   -1.5753  -5.7121  -4.4898  -4.3758  -0.6711  -5.8753   3.2373  -0.0331
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]