Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -1.3509   2.1477   6.6305 -12.6714  16.2506  -8.0182  -8.5878   1.6294
#>   -0.8889   0.9281   0.8775   8.4380  13.4555  -2.2574  -7.4528   8.4704
#>    2.4296  -6.5054   0.0372  -0.3790  11.1208  -1.3698  -5.4003   2.3944
#>   -2.8202   1.2180  13.6962  -3.3624 -11.6597   9.8679   9.8722  19.3768
#>   -2.3598  -2.7621   2.5804   1.3660   4.7200  -0.8378  -0.2113  -4.1866
#>   -4.0555   7.5137  -9.1743  15.0477   5.1997  -5.7385   7.6663  -2.3539
#>   -2.1787   6.5116  -2.5161   7.2732   9.7471   7.2179   8.6798   9.8738
#>    1.3405  -3.8423   4.7501  -3.1847  12.4428  16.6958  -9.5555  -2.4811
#>   -2.7207   1.5372  -0.0184  -6.3958 -10.1509   2.6792   8.5617  12.9956
#>    3.4383   0.2126  -0.7519   7.9839   2.6097   3.1415   8.8253  -6.2291
#>   -4.8263  -0.0440   6.5269  13.5835  18.6588   4.9789  -7.1498 -17.3985
#>    1.6681  -3.0713   0.2717  -3.7636  -4.3060 -14.6431 -12.8947  -3.5762
#>   -5.2408   6.7458   3.8928  -7.1140   2.6091 -14.9118   5.9670  -8.0643
#>   -0.3515   0.2050   2.0208   3.8218  -5.9547  -1.1384  -5.8426  -3.4869
#>   -3.5587   6.2850  12.4880  -9.9811 -14.0119   6.6712  -7.4807   6.9657
#>   -0.5432  -0.1759   8.4913 -10.2947  -7.2401   4.5054  14.4794   2.7666
#>   -5.1274  -4.1564  12.5711  -1.5259   9.1220  22.0858  -6.3091 -10.4710
#>    4.1100 -11.6122  -2.1532  -1.8814  -6.1435  -2.7956   5.0119 -13.9556
#>   -2.0880   0.0510   9.6320   5.1290  24.8029 -16.8858   2.6140   0.9398
#>   -1.9233   2.4477  14.0269   2.5987   1.1982  -6.2534  -8.2137  -1.0270
#>    5.9314   1.0005  15.8520 -13.3751   0.1466   8.6441  -4.8715  15.3322
#>    1.4007   1.8347   1.1910   4.0954   7.2071  -3.4070  -1.8618  -4.0433
#>    3.6727  -4.6370   7.5509   4.5975  -7.5279   5.9560 -12.8384  -1.2595
#>    0.1632   4.8333   1.0991   6.1375   9.4652  -1.3426 -12.0062   1.9991
#>   -0.9556  -3.6237  -2.8699   6.0835   9.6401  -9.5765  -3.2611  -4.1138
#>   -3.0628   7.4970   3.9260   8.0829  -1.3891  12.3683   5.1634  15.5280
#>    1.0777  -7.2460   2.7012  -6.5287 -14.8103  -5.2588   2.5157   7.9205
#>    3.3560   6.9010 -16.2793  -2.9881  -3.3136   2.2957  -3.6881   9.6563
#>    3.0842  -3.5241  -6.8094  14.0918   9.2812 -19.5712  -4.7379  -4.6223
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]