Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) =  Columns 1 to 6  5.5824e-01  1.5568e+00 -3.5899e+00  4.8409e+00 -5.1523e+00 -1.0610e+01 -3.6073e+00  8.2626e+00 -1.0460e+01 -4.5838e+00  1.3538e+01 -1.9043e+01  3.5589e+00  1.4752e-01  2.6766e+00  2.8268e+00  5.8242e-01  1.2035e+01 -6.5997e+00  4.2401e+00  2.9220e+00 -7.7766e+00  6.6885e+00 -8.8617e+00  2.2423e+00 -1.3624e+00  6.8611e+00  6.7216e+00 -2.0498e+00  9.2312e-01  1.1163e+00 -1.6133e+00 -6.9466e-01 -1.5299e+00  4.7933e+00  1.0029e+01 -1.0979e+00 -3.9101e+00  4.6418e-01 -7.4945e+00  3.0861e-01 -4.1607e+00 -1.8020e+00 -8.2219e+00 -2.2464e+00  1.1125e+01  3.5244e+00 -1.0446e+01  8.5263e-01 -2.7157e+00 -3.4919e-02 -1.4608e+01 -6.0263e+00 -4.2219e+00 -1.2968e+00 -1.0470e+00  7.9247e+00 -6.3273e+00 -3.3719e+00  1.4491e+01 -6.5035e+00 -6.7635e-01  4.0060e-02 -1.1969e+01  3.6076e+00  5.2192e+00 -2.2571e+00 -1.4954e+00 -7.0076e+00 -1.6478e+01  3.0080e+00 -1.3117e+01 -4.9779e+00 -6.9635e+00 -1.9183e-01  1.8381e+01 -6.8501e-02 -3.3879e+00 -5.3211e+00 -6.5664e+00  1.3557e+00  1.7008e+00  8.0936e+00  3.1890e+00 -2.3469e+00 -5.4730e+00  1.4445e-01 -3.5752e+00  1.0532e+00 -6.8081e+00 -1.6722e+00  2.4545e+00 -7.5779e+00  2.4759e+00  4.9605e+00  9.8516e+00  1.7924e+00  6.0222e+00 -9.9743e-01  3.9211e+00 -3.9162e+00  7.4303e+00 -1.6796e+00  9.0835e-01 -1.8040e+00 -2.6822e+00  6.0396e+00  3.1100e+00 -7.2146e+00 -7.5007e+00  6.8544e+00 -1.0064e+00  2.5050e+00  4.7136e+00 -2.8367e+00  2.6611e+00 -2.1743e+00 -8.1938e+00  5.5760e+00  6.6281e+00 -1.9880e+00 -8.0020e-01  1.9935e+00  1.2648e-01  2.1877e+01 -2.9910e+00  4.3924e+00 -1.2231e+00  9.3464e-01  5.8710e+00 -8.7310e+00  6.2138e+00 -2.2255e+00 -1.1204e+00  1.0021e+01  1.4431e+00  4.3682e+00  9.2798e+00 -5.3032e+00 -5.4961e+00  5.1998e+00 -2.1358e+00 -2.1632e+00 -2.4985e+00  3.5438e-01 -7.0084e+00 -4.6502e+00 -6.3219e+00 -5.1043e+00 -7.2658e-01 -5.1533e+00  8.1571e+00  1.6982e+00  5.9278e+00  1.4608e+01  9.4368e+00  5.4607e+00  9.7044e+00  5.5784e+00  4.5008e+00 -5.7411e+00 -6.7215e+00 -7.4817e+00  7.5307e-02 -9.4394e+00 -4.0735e+00  3.9900e+00  1.3420e+01 -9.1760e+00  2.2771e+00 -3.5968e+00 -4.2068e+00  2.8682e+00 -1.8024e+00 -4.9702e+00 -2.0279e+00 -4.1338e+00  1.7648e+00 -2.5170e+00  6.1837e+00... [the output was truncated (use n=-1 to disable)]
#>