Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8  -2.7603  -7.6490  -1.5395  -1.2865   0.1603  -4.3927   1.8567  -7.5046
#>   -4.4104  -1.1082   2.7405   4.7012   9.2401   6.3444  -6.3221   4.4170
#>   -4.6426   6.7129   0.7837 -16.1319  19.6525   8.4409   4.1343 -11.6119
#>    1.2809   0.3172  -8.7457   5.5598   0.2928  -1.9854  -8.9629  -3.9646
#>   -4.2911   1.6112   1.2141 -14.3788  10.9583  -5.5555   4.2494  -9.8882
#>    6.7736   1.8202   5.9378  18.2255  -5.7694 -19.0071  15.4113  -2.7397
#>   -0.1535   4.7522   1.5465 -12.0990   3.0110  10.1040 -10.6089  -6.5958
#>    0.5421   0.0903   2.1985   0.1613   1.9835  -8.1973   8.7975 -11.4114
#>    2.2028   3.4921  -1.9965 -31.5730  -0.6418  21.2707   9.7914  -4.6150
#>   -6.6771   7.8494   1.5740 -11.9789   0.5260  -2.0458   1.4243  -7.4406
#>   -3.1098   0.9808   6.8614 -23.9913 -11.4040  10.9254  -4.4803  -9.9509
#>    0.9574   2.9532   8.8563   2.6576   1.4444   0.3824 -10.6328  -8.9587
#>    2.5785   6.8731  -7.3122  -3.2582   4.6097   4.2349   3.6640  -1.0981
#>    0.1311   1.2819  -1.6007   8.7456  24.9043 -16.8731  -7.1150   8.2227
#>   -1.6632  -5.8502  -8.1136   3.4923   4.2641   6.6661  -6.4979 -14.0446
#>   -1.8965  -0.6090   5.0421  -5.8326   6.7916   0.1923   6.8011   5.3724
#>   -0.4575  -4.3534  -5.3142   8.0256  18.4511   2.1460  -5.1669  11.3640
#>   -5.8308  -4.4976   1.3206  -0.2379  -6.5679   8.6821  -7.2679 -14.0794
#>    5.9534   5.1067   3.7950   8.6273  21.9407  -2.5103  14.7087   3.1885
#>   -1.1842  -4.4865   1.4489   5.5162  -1.1731 -12.8710   6.5403  14.9231
#>   -7.2479  -2.0468   9.8004   5.7572  -5.9115 -10.5602  17.2185   7.0845
#>   -8.1645  -2.5499   1.0236   6.2030  -5.0975  -0.8134   7.5033  -5.9634
#>    1.6056   0.5862   5.7314   4.5280   3.7145  -5.5599  23.8166  -1.1289
#>    8.3585   0.0623 -10.9525  11.5875  -8.4416   3.1959  -5.8492  -8.7686
#>    3.4150   7.4765  11.4387   4.5895 -13.7320   5.7826  -0.0703  -6.8414
#>    0.6831   1.6184   2.7114  -4.9809   2.3055  11.8984  -4.9455   0.3555
#>   -5.7330   0.6888   3.3260 -13.4284  14.1861  -5.5951  -4.6493  -6.1724
#>    3.5035   5.9898   0.8071  -3.0411  15.4313  17.8105  -9.0091  -0.8675
#>   -6.4143  -0.9985   6.0080  15.2094 -12.6596 -13.0493  20.2408   8.5632
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]