Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) =  Columns 1 to 6 -4.7006e+00 -5.2750e+00  4.2207e+00  1.2433e+01  5.7151e+00 -7.8618e+00 -8.3766e+00 -4.9990e-02  1.4273e+01 -3.9222e+00 -1.3761e+01  7.1088e+00 -8.2017e-01 -1.6586e+00  2.6799e+00 -7.4118e-01 -5.2599e-01 -6.7811e+00 -6.1286e+00  2.9250e+00 -6.0888e-01  2.8342e+00 -2.4654e-01 -3.0206e+00  3.0731e+00 -5.1014e+00  7.3588e-01  2.1214e+01 -8.4973e+00 -1.4162e+01 -3.2838e+00 -3.4990e+00 -8.6606e+00  4.4556e+00 -3.6351e-01  1.5624e+00  3.6725e+00  8.1113e-01 -3.0498e+00  1.0765e+01 -1.9198e+00 -7.7586e+00  4.1162e+00 -9.9898e+00 -9.7567e+00  3.7782e+00 -1.2708e+01  4.3813e+00 -3.3597e+00 -4.2322e+00  5.0340e+00  4.3010e+00  1.5039e+00 -1.1339e+01 -3.1424e-01  8.5561e+00 -5.7266e-02 -1.2872e+01 -7.6437e+00  6.9985e+00  2.5881e+00  6.6676e+00 -7.4260e+00  9.7313e+00  1.2995e+01 -1.1118e+01  3.9452e+00 -1.6648e+00 -5.9042e+00 -1.5555e+01  1.0095e+01  1.5421e+01 -5.1210e+00 -1.5451e+01  6.7177e+00  1.5234e+01 -1.1986e+01  1.0175e+01  6.6644e+00 -6.9758e+00 -5.8333e+00  1.7639e+01  4.1022e+00  1.3620e+01 -3.7393e+00 -4.0552e-01  9.7948e+00 -6.5392e+00  7.4005e-01  1.6550e+00 -1.6155e+00 -2.7484e-01  9.0856e+00 -1.3557e+00  6.9630e+00  6.7109e+00 -1.4414e+00 -5.0090e+00  1.0697e+01  5.6741e+00  1.1188e+00  8.5970e+00  1.3772e+00  1.9174e-01 -1.4578e+00 -7.4402e+00 -1.3735e+01 -2.4023e-01 -1.6020e-01  2.8091e+00 -8.8331e-02  8.4117e+00 -3.6362e+00  2.3877e+01  2.0931e+00 -8.1784e-01 -4.3453e+00  9.7732e-01  2.5429e+00  7.9237e+00 -7.7054e-01 -1.3973e-01  1.4395e+01 -1.6429e+01 -3.3858e+00  3.2402e+00  4.0933e+00  5.6089e+00 -1.3750e+01 -4.0227e+00 -9.4949e+00 -1.2093e+01 -6.4772e-03  9.2663e+00 -2.1471e+00 -1.6027e+01 -2.5951e+00  2.3630e+01  1.7412e+00 -2.6755e+00 -7.4278e+00 -1.3143e+00  5.8853e+00  1.2474e+01 -4.2200e-02 -6.6699e+00 -3.6603e+00 -1.1160e+01  3.1322e+00  1.0866e+01 -1.9963e-03 -6.5937e-01  9.2514e+00 -6.9742e+00 -3.2345e-01 -5.6419e+00  3.9430e+00 -4.5381e+00 -1.0058e+01  3.1475e+01 -8.1016e-01 -7.3532e+00  8.4739e-01  5.4468e+00  5.5064e+00 -7.6597e+00 -2.3186e+00 -1.0491e+01 -6.8537e+00  1.9558e+00 -9.5197e-02 -1.9329e+00 -3.2018e+00  9.2094e+00  7.0755e+00 -2.4706e+00 -5.4188e+00 -5.6548e+00  8.5438e+00  7.1449e+00... [the output was truncated (use n=-1 to disable)]
#>