Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   6.1622   3.0141  -0.9979  -1.6551   6.4761  -5.7683   0.2416 -11.9482
#>   -2.2989   3.3845  -8.6105   0.0549  -9.2399  18.3673   2.8969  15.5643
#>   -1.0452   6.0094 -11.1807  -0.1301  -4.4585  12.5272  -9.8602   2.5989
#>   -5.2541  -3.3215   1.4680  -0.2534 -12.0495   3.1877  -2.4672   8.7377
#>   -1.8018   4.4094   0.3353  -1.5305  -0.4717   2.5462 -11.5128  12.2849
#>    0.1210  -2.0106  -9.9963   1.0783   2.0353  13.7555  11.5960   7.1225
#>   -3.1924   4.0109  -4.7015 -14.2118  -9.0681  -8.6766   8.8016  -5.2652
#>   -4.3021  -3.5989   2.7651   7.8082  -1.6216  -6.0497   3.6556 -11.6530
#>    2.3110  -0.5832  -9.0004   9.5121  -1.9636   2.0894   4.1858   8.6954
#>    3.6691  -4.9328  -1.0222   2.2166   2.3011  10.8841  -7.4093  26.9730
#>    1.9090   0.0658   3.8506   3.0809   0.4697  -8.3507  -2.5357  -1.3697
#>   -1.6854   4.7432  -0.2102  -8.5356   3.1306  -8.8324 -14.1569  -6.6755
#>    0.6871  -2.6258   0.4541  18.7418  -2.7959  -6.6766  -9.1449   1.3552
#>   -1.6638   3.0591   9.4624  -4.2931   8.3756 -14.9954   7.6996   1.9334
#>    7.3743   6.5332  -7.3059  13.5625  -0.7170  13.3962  -4.9634  -6.5921
#>   -2.4535  -6.6049   3.6158 -12.9725   4.9004  -0.3311 -10.6670  10.4100
#>   -1.1893   0.6519   0.1278  -3.6227   9.9430   8.8949  -0.0144   6.8892
#>   -1.0372  -1.1275  17.8376   8.7933   3.9011  -4.1269  -3.5186   1.3628
#>   -2.3129  -4.7976  -1.2249  -0.0695 -13.3164  -2.8195  -5.6357  15.6292
#>    1.2428   5.0103   6.1601   0.7839  -4.4143 -10.1713  -5.4127   2.3110
#>   -1.3074 -14.2119  14.8997 -11.1018   6.3570  -1.0411 -13.9191   2.3714
#>    0.6362   1.0701  -7.9386   9.9903   4.8885  -6.6783  12.0296 -17.6702
#>    4.0057  -5.2909  -8.1610  -9.0322  -0.5221  -1.1484  -3.9342  -0.6423
#>   -0.8273  -5.5577  -5.0385 -12.6608   4.7167   4.8329  10.6469   0.9087
#>   -1.4550  -4.1424   4.8476 -10.4040  14.2188   9.4972   8.5873  -0.6871
#>   -0.7951  11.1323  -4.6557   7.5739   6.6803  10.5051  -1.1949   1.0344
#>    2.8080  -2.0878   7.3263 -14.9923   2.6639   2.7362 -13.9002  -2.4684
#>    2.8635  -0.4150   4.1800   7.3295  -5.1352   5.3667  14.7919  16.9246
#>    1.5430   1.8389  -4.6813  -1.1657  -6.6271 -11.1413   7.8052 -11.8207
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]