Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6  7.4362e+00 -1.0006e+01  1.1168e+00  4.9690e+00 -5.8433e+00 -9.8683e-01
#>  -3.1772e+00  3.4733e+00 -9.9185e+00 -3.6950e+00  2.9505e-01  7.9996e+00
#>   1.3081e+01  4.6162e+00 -2.9053e-01 -1.3031e+01  6.6212e+00 -3.8164e+00
#>   2.9257e+00 -1.1461e+01 -4.2721e+00 -7.1295e+00  6.9823e+00 -2.9571e+00
#>  -2.6087e+00  1.0373e+01  5.3542e+00 -1.0568e+01 -2.2241e+01  1.7815e+00
#>   1.0932e+00 -9.5887e+00  3.6332e+00  3.1038e+00  5.2959e+00  7.6827e+00
#>   2.2437e+00 -1.3959e+01 -8.1075e+00 -5.1801e+00 -4.4649e+00 -1.2871e+01
#>  -6.1786e+00  2.3630e+00  3.1900e-01  3.1124e+00 -1.3159e+01  3.9692e-01
#>   9.8296e+00 -3.8939e+00 -3.8256e+00 -1.3196e+01 -5.0806e+00 -1.3847e+01
#>  -2.7879e+00  1.3198e+01  1.6459e+00  2.5276e+00  1.7264e+00  8.9373e+00
#>   2.4473e-01 -4.1496e+00  3.6370e+00 -1.1755e+01 -1.2381e+01  3.1800e+00
#>  -8.3979e+00  5.8234e+00 -4.7637e+00  4.8313e+00  1.2093e+01 -1.0094e+01
#>   3.1202e+00 -6.8924e+00  5.0364e+00  3.6972e+00  3.5331e+00 -1.3514e+01
#>  -1.7707e+00  2.4737e+00  7.3336e+00  9.9063e-01 -2.0168e+01 -1.8768e+01
#>  -1.1676e+01  1.6941e+00  1.4010e+00  6.5276e+00  5.9614e+00  6.9270e-03
#>  -5.9908e+00 -7.4142e+00 -3.6286e+00  4.9738e+00 -7.2187e+00  7.7983e+00
#>   9.3248e+00 -8.9412e+00  4.7828e+00  4.1770e-01  1.6196e+01  6.2733e+00
#>  -1.7275e+00 -2.9394e+00 -8.9101e+00  1.3272e+01  1.0457e+01  4.1500e+00
#>  -2.5478e+00 -4.7998e+00 -4.1689e+00 -1.9166e+01 -1.0616e+00 -4.6521e+00
#>   2.0050e+00  9.9641e+00  5.2962e+00 -6.2119e-01 -1.0690e+01 -4.7125e+00
#>  -8.2638e+00  5.3833e+00  1.7161e+00 -4.3633e+00 -5.2666e+00  6.8849e+00
#>   2.2545e-01 -3.9529e+00 -1.2192e+01 -3.0525e+00 -4.4056e+00  1.0976e+01
#>   2.6789e-01 -3.1142e+00 -7.8347e-01  7.2189e+00  1.0246e+01  2.8891e+00
#>  -5.8938e-01  4.5836e+00  2.5912e-02  8.8011e+00 -4.0650e+00  1.1329e+01
#>  -2.0624e+00 -2.3867e+00  8.1108e-01  4.4705e+00 -7.7347e+00  2.3757e+00
#>   8.9801e+00 -1.2866e+00  3.1426e+00 -1.3198e+00 -8.2487e+00 -5.0701e+00
#>  -7.1906e+00 -5.0979e+00 -9.9125e-01 -9.1305e-01 -6.0246e+00 -1.4134e+01
#>  -3.3704e+00 -1.2427e+00  4.0257e+00 -6.3462e+00 -3.0292e+00  6.6130e-01
#>  -9.2219e+00 -6.0051e+00  5.9188e+00  1.2507e+01 -6.1914e+00 -2.1726e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]