Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 6 -6.7746e+00 -8.1241e-01 -4.4340e+00 -3.0766e+00 -1.6681e+01  2.4573e+00
#>  -2.3960e+00 -3.5930e+00 -1.2448e+01  2.4147e+00  1.1164e+00  1.0814e+01
#>   6.2391e+00 -4.7095e-01  1.2573e+00  4.7812e+00  1.1365e+01  2.6605e+00
#>  -2.5897e+00  1.2307e+00 -7.3395e+00  1.0240e+01 -1.0091e+00 -1.6647e+00
#>  -6.3255e+00  3.7988e+00 -2.9435e+00 -7.3505e+00 -5.5807e+00  5.4028e+00
#>  -5.1865e+00 -3.3282e+00  1.2954e+01  4.4636e+00  8.5784e-01  8.7089e+00
#>   8.6812e-01  3.2896e+00 -1.1082e+01  5.3243e+00 -4.9157e+00 -7.5853e+00
#>   7.6040e-02  7.2306e+00  6.8061e-01  8.1047e+00  3.5795e+00  8.0962e+00
#>   4.8451e+00  2.3722e-01 -3.7186e+00 -9.7677e+00 -2.2154e+01  5.7878e+00
#>   5.0557e-01 -2.0678e+00  1.0018e+01  1.3372e+01  4.0209e+00 -2.4922e+00
#>  -5.3960e-01  7.2940e+00 -2.9061e+00 -9.4397e+00 -3.3270e+00  5.2779e+00
#>   1.3113e-01 -4.0231e+00 -7.9611e+00 -4.2350e+00  2.8343e+00 -1.3757e+00
#>   6.0677e-01 -2.2526e+00  9.7778e-01  1.3570e+01 -8.1484e+00  1.4348e+01
#>   6.2802e-01 -1.1771e+01  4.1464e+00 -5.0732e+00  2.6240e+00  1.3457e+00
#>   4.5481e+00 -6.0094e-01  1.5565e+01 -7.7423e+00 -1.7338e+00 -8.4230e+00
#>   1.6357e+00 -8.4276e+00  5.7280e+00 -1.1550e+01  1.0987e+00  6.0091e+00
#>   2.7374e+00 -1.2366e+00  1.5912e+00  1.2489e+01  9.5057e-01  1.7753e-01
#>  -3.3513e+00  1.4449e+00  7.2598e+00 -3.5278e+00 -4.6190e-01 -7.2708e+00
#>  -4.3150e+00 -1.4040e+00  4.8618e+00  4.6452e+00  1.1516e+00 -1.7588e+00
#>   1.3029e+00 -3.3646e+00  3.8172e+00  1.6457e+00 -5.7519e+00  1.0257e+01
#>   2.9155e+00 -1.0793e+01 -8.6930e+00  7.9348e+00  1.0054e+01  5.6242e-01
#>  -8.5737e-01 -4.2122e+00  2.9005e+00 -1.3667e+01 -1.8301e+01  8.0919e-01
#>  -3.8778e+00 -7.3393e+00 -3.0016e+00  3.7547e+00 -9.1400e-01  1.0143e+01
#>  -2.3662e+00  1.1028e+00  4.9092e+00 -1.9712e+01 -9.6386e+00  8.8875e+00
#>   5.7273e+00 -3.3208e-01  4.5091e+00 -7.1155e+00  4.5085e-01 -1.7378e+01
#>   5.2725e+00  8.1924e-01 -2.3901e+00  9.1869e+00 -1.6481e+01 -1.1055e+01
#>  -5.8471e+00 -4.0000e+00  7.9884e+00  9.2609e+00  1.2046e+01  5.4402e+00
#>  -3.2544e+00  2.7291e-01 -5.0235e+00 -3.3138e+00 -6.1394e-01 -5.9341e+00
#>   3.4232e+00  7.9690e-01 -8.1593e+00  3.6106e+00  6.2611e+00 -1.6794e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]