Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   0.2941  -5.0772   1.6125   3.8268 -12.5572  -6.8332  10.3453  -3.5113
#>   -0.8774   9.6810  -3.3659  -1.7304   0.6027   4.2310   1.4705  -3.6962
#>   -0.5969  -7.9692   5.4336  14.9713  17.8538 -18.6989  -2.9870   7.7017
#>   -6.8205   7.9288   3.0663  -2.4659   2.8688  -3.3221  -2.8244   2.1673
#>    8.2921  -0.5969  -8.1048 -10.7060   3.1811  -1.5150  -0.2334  -8.4774
#>   -0.6128   7.7357  -8.3314   7.9066  -6.2290  10.7602  -1.7605  -7.5379
#>    1.7140   4.8625  12.0464   9.4346   7.6970 -11.8430  -4.6654   1.5184
#>   -5.6557  11.7680   0.2584  -4.7230   9.5670  -6.6502   2.2354 -11.8219
#>    9.3802  -5.0782   4.1571  -5.7680  20.4820  -2.3154  22.0551  -0.3006
#>    0.4918   4.3004  -5.9645   7.4348  -5.5768   6.4858 -14.3108   9.7059
#>    0.5983  10.5583  -8.8940  -3.5170   5.8048  -1.3321  -9.6006  -6.0427
#>    4.8436   3.4946   3.8251  -6.4758  -6.8748 -12.9379   0.5731   1.7257
#>   -3.4143  13.5077  -4.2012   0.0869  -8.1866   9.1363   8.5425   7.5311
#>   -2.7695  -6.5825   1.0577   3.2935   8.2997  -9.2406  -0.5059 -14.6475
#>   -4.8099   2.5654  -3.9192 -15.9955  -8.4412  -0.6823 -10.3250   4.8686
#>    1.9628   5.9839  -8.2658  -0.3582   8.2467  -2.7752  -3.9058  -7.8915
#>   -0.8386  -5.2409   0.6157  -8.4723  -6.7347  -1.8793  -9.9669 -11.3819
#>    1.2858   5.8777  -1.5287   8.6779   8.7435   2.1554  -8.1258   4.2092
#>   -4.2713  -5.3007  -0.5005   1.0536   9.9791 -20.0750 -15.2558  16.6823
#>    1.1937 -13.9071   0.1890  -3.2194   6.6973  -5.8448   4.0622   1.4531
#>   -0.3757   7.8491  -4.0428   8.7549  11.0493  -5.8511  -7.4438  10.2607
#>   -4.2617   3.8914  17.2463 -17.1492  -4.6334   5.8008  -8.1465 -15.5788
#>   -1.6917  -2.0005   0.2993  -9.5672   0.4327   1.3786   5.4902 -17.4087
#>   -6.3144  11.3869  -7.3429  -8.1201 -13.2243   9.7718   2.2884  11.9298
#>    2.9204  -3.6797   1.4260   1.9340   3.0551  13.0020   0.8002  -8.4526
#>   -4.7789  -8.5774   1.2436 -14.8198 -10.5065 -11.2886   6.1707   2.0392
#>    2.4979  -4.5571  -0.3451  -6.0607 -14.5953   3.0561  11.8624   2.1289
#>    0.5688   2.6199  15.5711  -5.8390   1.9992  -4.6730   8.5897 -24.3284
#>    0.4859  -4.2466  -6.2973  12.3988  13.8836  -2.9904   2.3680  -2.6464
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]