Conv_transpose1d

Conv_transpose1d

Usage

torch_conv_transpose1d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sW,). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padW,). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dW,). Default: 1

conv_transpose1d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 1D transposed convolution operator over an input signal composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose1d() for details and output shape.

Examples

if (torch_is_installed()) {

inputs = torch_randn(c(20, 16, 50))
weights = torch_randn(c(16, 33, 5))
nnf_conv_transpose1d(inputs, weights)
}
#> torch_tensor
#> (1,.,.) = 
#>  Columns 1 to 8   3.4089   1.9923   8.3964  -1.8326   6.8711  -5.1467  11.5144   4.3971
#>   -2.7702   5.5465  -0.1016   6.9843 -15.7568   5.0906   7.5009   4.4790
#>   -3.6100  -6.2406  11.2735 -11.8163  -7.3309  -0.0506   5.0156  -1.1719
#>    6.7336   0.9274  -0.4033  -5.0775   1.3114  -0.7929  -8.7918   6.4456
#>    1.6161  12.6772 -11.7664  -2.5311  -5.1420  -0.1793   4.1697  -5.7172
#>   -0.4337   3.0913   3.7897   2.6436  -2.6834  11.0540 -10.4828  -6.1375
#>    4.9754  -0.9639  -8.6367  -0.0675  14.2478  -1.2999  -6.0835   5.1482
#>   -4.7714   7.6681   8.3212  -4.0659   7.0583  -4.9608   7.8702  10.5812
#>   -3.4760   9.5754   3.5962  12.4364 -11.5956 -14.3728   5.2243   9.4770
#>    4.8543   7.9437  -3.2524 -11.6211   2.6143  -1.8475   5.7013  -6.5386
#>   -9.2844 -10.4435   0.1003   4.3454  -7.5814  14.1377  -0.0151  -3.6578
#>   -2.0506  -8.9131 -15.2067  14.5911  11.7874  -0.2957  -8.9815 -15.6720
#>   -4.7661   1.6460  -8.9024  14.4496   7.2314   3.8777   3.8603   9.4748
#>   -4.4931   7.0800  -3.3859   8.3677  -7.7306 -11.3296   9.4916  10.3365
#>   -0.5200   9.9860  10.6308  -2.7165  -8.2239  17.2875   1.5051   8.1951
#>   -1.1093   0.3947   8.1787   3.7629   6.4310  -6.2769   7.1425  -3.3923
#>    4.4558  -1.3393  -8.7477  -1.8713  -3.7511 -13.2252   4.6304  19.0653
#>   -1.0424  -3.0065  -2.0183   2.3221   9.5649  -5.8569  10.2109   6.1251
#>   -1.1786   3.9590 -13.8480  -3.9788  14.2736   1.9588 -23.2189 -13.5426
#>   -5.0767  -7.4101  -6.3379  19.0710  -7.9112  10.5089 -16.1527   5.0957
#>   -1.2417  -0.4376  -0.8849   2.1248   2.9225  -5.3589  -9.6065  17.4514
#>    0.4635  -4.0378   1.6771   5.9117 -14.4362   1.7423  -1.0227  -1.0648
#>    4.1225  -8.6344  -0.9898  16.8196  11.2528  -2.7512  -7.3270  -5.0332
#>    0.1342   6.5893  -1.5447  18.3998  -5.7629   1.6465 -18.5898  -2.5248
#>    5.1837   1.2207   2.5221  -4.9087  -1.0872   1.7922  -8.8508 -10.9090
#>    1.3307  -3.6528   5.1665   6.9779  -7.1443  26.0210   8.0342  -4.2664
#>   -2.5106   5.9716  -0.3857   1.9542  -4.5693   4.5889  17.5952   1.2114
#>    3.0610  -1.6509  -0.2090 -12.5600  -6.1317  -3.4714  -0.4867  -2.9715
#>   -0.7428  -6.9170  -2.7547  -7.3484  -1.1848   7.7382   5.5558   5.4516
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{20,33,54} ]