Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -11.6503   0.9092   0.8537  11.6027  -4.0104
#>    4.9350   1.3871  -1.8473  -2.2584  -1.8432
#>   -2.4037  -0.3483  -3.1652   6.0101   5.5443
#>    2.6399  14.3157   4.4060   1.0629  -1.7242
#>    0.2730  -4.2674   0.2865   4.0603  -0.0955
#> 
#> (1,2,.,.) = 
#>  -9.3049 -7.5554 -4.6184 -2.0005  0.3483
#>  -3.7922  2.7088 -1.0215  6.7845  1.4839
#>  -0.2559  1.6758 -7.6997  9.8798  1.3343
#>   4.3250  6.2217 -2.9167  0.9186  3.5244
#>  -5.2581  9.8645  6.5341 -2.8123  0.6025
#> 
#> (1,3,.,.) = 
#>  -2.6867  3.6376 -8.0003  2.5752 -0.5381
#>  -5.3965  5.8811  5.8785  3.3568 -4.8534
#>   4.6565 -4.0644 -6.3777  0.5083  2.1373
#>   3.1525  4.2297  2.9901  3.3640  5.3181
#>  -3.0321  9.9206 -1.4412 -8.5012  6.7548
#> 
#> (1,4,.,.) = 
#>  -0.4869  0.8433 -4.6834 -9.0936  2.2643
#>   0.9137  3.5030 -5.8924  2.5013 -2.8058
#>   4.6447  5.7029 -2.8845  7.1939  0.5308
#>   4.9858 -0.8124  5.7068 -3.5370 -6.0685
#>  -9.8864  2.2727 -5.0801  3.2534  0.6895
#> 
#> (1,5,.,.) = 
#>   5.8999  4.1328  5.4112 -2.1028  1.2551
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]