Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   5.3295 -6.4927 -7.6208 -0.1141  0.7608
#>  -4.1880 -3.7116 -2.3313 -3.4566  5.3712
#>  -4.9034 -0.8722 -2.9895 -3.3059 -6.7131
#>  -6.2529 -6.5855  0.0686 -1.1436 -6.4516
#>  -1.1100  0.4901 -7.9923 -8.3011 -2.2420
#> 
#> (1,2,.,.) = 
#>   0.5145 -4.3450 -10.8106 -11.6117 -0.9510
#>  -1.8702 -6.8892  1.4450  5.5503  2.4024
#>   6.5994 -1.5248 -9.7187 -3.0634 -2.3928
#>  -1.7685 -4.4221 -1.7776  5.1650  4.1452
#>   4.6063 -0.1802 -2.8163 -5.3768 -7.6031
#> 
#> (1,3,.,.) = 
#>  -0.1259 -8.8426 -6.8760  2.5683 -0.4131
#>  -3.3883 -2.6068  6.3939 -7.0518 -10.6462
#>  -2.1024  0.5850 -3.8337  1.6976  0.4358
#>   0.4698  2.0318 -10.0762 -7.8092 -0.9082
#>  -3.5949 -4.5435 -2.8419 -2.1740  0.4248
#> 
#> (1,4,.,.) = 
#>  -5.5781 -4.4863 -5.9124 -0.4392 -4.4891
#>   4.6254 -8.6013 -8.1149 -5.6861  0.8627
#>   2.2095 -3.0356 -7.4547 -9.1830  0.4957
#>  -7.2862 -6.9327  7.1909  1.3707 -0.9293
#>  -2.6224 -5.5069 -2.4534  0.1845  3.8667
#> 
#> (1,5,.,.) = 
#>  -5.9701 -4.3750 -0.7813 -4.3618 -3.2274
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]