Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    3.5900   6.4060  11.4520  -0.0769  -3.8879
#>   -0.2932  -5.9292 -11.1706  -8.3051  -3.7300
#>   -1.8497  10.2684   7.6160   0.8416   1.5238
#>    1.6493   1.7143   5.9357   5.6548   5.5937
#>    2.5709   0.4332  -3.5398  -0.8920   0.6443
#> 
#> (1,2,.,.) = 
#>   6.9935 -7.7474  2.0848  9.2413 -5.5207
#>   5.2778  2.5618  1.2869  1.8082 -6.7645
#>   2.0996 -2.8223 -0.6235 -10.8965  0.8756
#>  -4.4765  0.1286 -1.0530 -0.1431 -5.6328
#>  -5.7503 -0.0773 -1.6279  3.2790  0.6087
#> 
#> (1,3,.,.) = 
#>    2.6716   2.0686   4.3071   3.3022  -7.3272
#>   -1.9680  -8.5514  14.8744  -4.5575   1.9048
#>    0.3239   9.5901   0.1285   8.2762   3.3374
#>   -1.3412   6.5872  -0.5768  -5.4055   6.9406
#>    2.6775   2.7281  -8.4052  -3.3220   3.0406
#> 
#> (1,4,.,.) = 
#>   -5.7062  -4.1822  10.0604  -1.9627   6.5538
#>   -0.7636   2.2275  -0.4996  -7.0022  -0.0724
#>   -1.0028  -6.3092  -7.9503  -6.3829   0.1998
#>   -0.9848   4.1800   2.8604  -2.1550  -0.3843
#>    3.5105  -0.1536   3.3530   4.7807  -1.1392
#> 
#> (1,5,.,.) = 
#>  -2.0011 -4.4034  1.4732  3.1983  3.6007
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]