Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    5.4706  -1.7291  -4.5292  -2.8441  -5.1721
#>   -1.2710   4.2358  -0.6962  -5.4919  -1.4708
#>   -5.6899  -2.1581  -7.7634   3.7047  -1.6576
#>   -4.9140   3.3940  -2.0245  11.3940   1.8974
#>   -5.3683  -9.6487   0.2460  -2.6353   3.2006
#> 
#> (1,2,.,.) = 
#>    4.2951  -8.3126 -14.3170  -7.5075   6.7367
#>  -10.6025   7.0583  11.4378   7.7614   1.5503
#>   -3.7719   5.4981  -6.8145   3.2533   1.3907
#>    3.8941   5.6589   3.7745  -1.7770  -5.3072
#>    4.6969   1.1324   8.4419   5.5210   4.6490
#> 
#> (1,3,.,.) = 
#>   4.4916  2.5967 -6.9736  0.0763  0.5071
#>  -7.6495 -6.5821 -3.1312  2.8986 -3.7167
#>   2.3058  4.2057 -5.5546 -1.2488 -0.7592
#>  -5.1025 -1.6230  2.0158  2.1622  0.0640
#>  -1.4962 -2.8349 -8.8866 -3.4539  0.9469
#> 
#> (1,4,.,.) = 
#>  -2.5069 -1.8034 -0.9695  4.5808 -0.4410
#>   0.7279 -3.5670  4.3968  6.5522  3.7686
#>   2.5786  9.3008  7.3970  4.6523  1.5671
#>  -3.6734 -4.0606 -5.1933  2.7787  0.6971
#>   0.0978 -0.4877  2.1307 -2.2919 -6.1639
#> 
#> (1,5,.,.) = 
#>   -2.5355  -2.9037   5.7243   5.1620   9.9214
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]