Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -1.6821  2.6414  1.2534 -0.7151 -0.9242
#>   1.1280  1.9635 -0.2283 -1.6827 -0.7778
#>  -3.2615  0.1517  7.1029  5.5545 -0.9781
#>  -2.1206  0.6832  2.8834  5.1569  1.8530
#>   0.0967  3.0429 -0.7379  1.7419 -2.0079
#> 
#> (1,2,.,.) = 
#>   2.2424  4.6324  7.8652 -3.2356 -2.9238
#>   1.1887  0.1157 -5.4252  7.2916 -2.8776
#>  -7.2607 -3.8823  7.4114  7.3366  0.8298
#>  -5.3161 -7.9274 -2.1592  2.9767  2.3152
#>   7.8468  7.9571  4.7943 -0.3054 -0.1074
#> 
#> (1,3,.,.) = 
#>   6.3327  3.5469  1.0530 -1.8972  0.3989
#>   4.3065 -5.9783  5.3663 -5.0393 -5.1756
#>  -3.9558  0.7816  8.9839 -0.1187 -7.2703
#>  -9.0998  6.2627  4.3337 -3.2564  2.2123
#>  -3.3697  4.2838  9.8259  6.9328 -3.8594
#> 
#> (1,4,.,.) = 
#>   -1.5946   5.8629  -0.8543  -2.0388   1.7865
#>    7.3287   2.7593  -0.3017  10.0918  -1.3019
#>    0.6938 -10.5804   0.1996   3.4778   7.0289
#>   -0.6686   3.8979   2.9748  -9.4749  -5.9675
#>   -3.8092  -0.2089   5.6509   8.4628   4.0846
#> 
#> (1,5,.,.) = 
#>  -1.4298  0.4295  8.6266  6.1454 -4.2817
#>   2.9544  7.2256  3.5424 -0.0213  4.4257
#>  -2.7603  3.6839  4.1262  6.0753  1.0348
#>  -2.7427 -6.2490 -11.2522  0.7969  0.0832
#>   6.9294  3.8769  5.1586 -4.4089 -6.1305
#> 
#> (1,6,.,.) = 
#>   0.6202 -9.3027  3.8024 -6.6818  4.3670
#>  -8.1638  0.5954  2.5451 -3.9621  2.5683
#>   5.8593  9.3100  1.0168  2.7279 -0.5912
#>   3.4269  0.0499 -3.1159 -4.6146  2.6742
#>  -0.4491 -4.7738 -7.8279 -2.1240 -2.4767
#> 
#> (1,7,.,.) = 
#>    6.6070   9.8367   3.1347 -11.2184  -3.6008
#>   -4.8907  -1.7500  15.3330   3.3529  -5.6560
#>   -2.3986   2.7287  -3.1102  -0.4348   0.5991
#>   -0.7044  -0.7790   3.9716   4.4696  -5.9365
#>   -1.5274  -1.1082  -0.2305   1.5522  -3.9049
#> 
#> (1,8,.,.) = 
#>  -3.3118 -3.2717  8.0907  4.9294 -3.5665
#>  -3.4650  2.4856  9.0280 -10.4059  1.7474
#>  -4.2311 -2.6608 -5.6510  5.4120  2.0459
#>   3.7805 -2.7066 -2.3894 -0.2429  2.4380
#>  -4.1844 -2.5535  4.2462  1.2168 -6.3516
#> [ CPUFloatType{1,8,5,5} ]