Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    2.1261  -2.5724   2.1759  -9.9047  -3.2362
#>   -2.4156  -5.1125   2.1613  -3.8496  -1.3240
#>   -6.2999  -5.8869   5.2565   2.6873  -1.5449
#>   -1.1748   4.8953   2.3580  11.5977  -3.2831
#>   -1.5455  -0.1162  -3.6726  -3.4606   0.8922
#> 
#> (1,2,.,.) = 
#>   -0.7875  -3.2915   9.4249  -3.8857  -5.2050
#>    0.6466   5.1890  10.0974  -0.4874 -10.6839
#>    2.7469   4.3057   4.6230   2.6352  -6.2175
#>    2.3930   4.0863  -5.4044   3.0437  10.9032
#>    4.5923   1.0201  -5.8029  12.6416   2.3233
#> 
#> (1,3,.,.) = 
#>   1.0050 -12.3128  1.3075  1.1700  0.8417
#>  -12.1977 -9.3306  2.3225  3.3178 -2.0384
#>  -5.1788 -3.3491  6.1495  9.4029 -2.2011
#>  -5.8191  2.3017 -2.7455 -1.7549 -1.3770
#>   1.2867 -1.1168 -1.8217 -1.4446  1.1745
#> 
#> (1,4,.,.) = 
#>   -3.5307   1.2268  -8.1419  12.8686  16.8372
#>   -6.9162   3.8113   5.3396  -8.4422  10.6286
#>    2.9289   7.3069   2.3129 -17.9609   0.2155
#>    3.3655   5.9098   6.6200  -8.2977  -0.7906
#>    2.0036  -2.8173   1.7280 -11.0928  -7.1110
#> 
#> (1,5,.,.) = 
#>   -1.3182  -1.6538  -0.7130   3.8332  10.8916
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]