Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   -3.6222  -7.9970   2.8105   3.1026   1.7725
#>   -4.7404  12.8775   1.0415  -0.8881   4.1451
#>   -2.2161  -0.0100  -7.8568   2.5352   7.7713
#>    1.3625  -3.1941  -6.9461   6.7935   2.9976
#>   -3.9660   0.9707  12.1132  -7.1898  -5.3734
#> 
#> (1,2,.,.) = 
#>   5.5750  4.9565 -1.2831 -9.3766  2.7933
#>   0.3677 -1.5841 -1.1948  2.0401  1.6757
#>  -1.1554 -7.9480  2.0611  3.8100 -2.3715
#>   3.6882  5.7345  7.8422 -10.5027  4.2258
#>   6.4450 -2.5137 -14.2397  6.3799  2.2110
#> 
#> (1,3,.,.) = 
#>   2.3055  1.7193  1.5106  5.7209 -0.4044
#>   5.4206 -5.3050 -1.7869  8.7362  4.7972
#>  -1.1797 -8.2598  0.6797 -5.7008  2.3267
#>   8.3444 -1.9313 -3.6077  1.1224  0.3162
#>  -0.1740 -4.8076 -2.2376 -3.3701 -7.0435
#> 
#> (1,4,.,.) = 
#>  -2.1321 -0.6399 -0.5702  6.2185 -6.6191
#>  -6.7085 -1.7849  0.0567  0.2547  4.4343
#>  -0.1791 -13.0960 -1.1711 -11.7881  7.9123
#>   1.6398 -8.7353  5.6190  2.7717  2.2932
#>   6.8436  5.9709  0.8596 -5.4111 -4.1101
#> 
#> (1,5,.,.) = 
#>    6.3795  12.9273   3.7923  -2.6331  -6.7760
#>   12.3485  -7.3105   3.6371  -0.5407  -1.6971
#>   -6.8453   8.1406   8.8767  -0.0699 -10.0042
#>    0.8129  -1.6930   3.5113  -0.9711   2.3083
#>   -1.1577  -2.1860  -1.9327   3.4738  -2.3536
#> 
#> (1,6,.,.) = 
#>    2.8265  -4.2878 -10.9216  -1.4770  -3.4568
#>   -8.9188  -7.0011  -0.9053  -2.3985  -0.2689
#>   -0.8018  12.4512   0.5999  -5.5394   1.6234
#>    4.3200 -10.4630  -4.2572   7.0144  -0.1696
#>   -4.6365  -0.5886   7.6997  -3.6797   2.4974
#> 
#> (1,7,.,.) = 
#>  -2.3134  1.9207  6.9010 -2.2058  2.9683
#>  -2.6449  5.9937 -0.7752 -2.3005 -1.7128
#>   3.1190 -11.4434  2.3856  3.4017 -0.6764
#>  -2.5913  6.5472 -0.4979  2.9762 -7.4267
#>   1.9456  3.5234 -6.3143 -3.6686 -0.6943
#> 
#> (1,8,.,.) = 
#>   -0.3478   4.7669   7.3665   6.3556  -2.5469
#>   14.4336   2.1104   2.8604   0.8508   7.2790
#>   -9.4439  -5.6837  -8.8314  10.9166   0.5517
#>   -8.0535   6.8200   4.4783  13.7326   3.1114
#>   -0.0064   8.4882  -0.5751   1.8667  -4.2371
#> [ CPUFloatType{1,8,5,5} ]