Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.3744  -4.0021   9.2813  -4.6533   2.1713
#>    0.0050  -6.7356  14.5738  -1.3197  -2.1540
#>   -8.3365  -1.2789  -1.8664  11.5486  -1.6233
#>   -1.3197   2.9443  -2.0605  -0.7448   3.1986
#>    1.9598   7.4299   5.3739   5.3108   2.2687
#> 
#> (1,2,.,.) = 
#>   10.7735  -5.8157  -8.6583   4.1100   0.4754
#>    9.3287   9.8687 -13.9147   5.6629  -4.2594
#>    4.4156  -0.0613  -3.2969   3.5086   1.2626
#>   -1.2804   5.6718  -6.6691  -0.6058   3.5967
#>    5.9743   1.2628   0.9414   7.0912   8.6674
#> 
#> (1,3,.,.) = 
#>   -4.6789   0.8959  -2.1885   2.0310  -0.4365
#>  -11.9505   5.5433   3.0480   1.8758   5.3260
#>    3.2911  -6.3856  12.0851  -2.1029   1.9040
#>   -4.4916   7.4859  -2.2141  -7.0288  -0.9222
#>   -0.2099   2.7998   4.0858  -4.6763  -1.1309
#> 
#> (1,4,.,.) = 
#>    4.0603  -2.1008   0.3741  -0.3874   0.0683
#>    2.8996   3.7472  -5.8523   2.9654  -0.3575
#>    0.5845   0.4552  -1.9663  -1.8625  13.6938
#>    1.8901  -0.7993  -3.2567   8.2104  -3.9014
#>    6.3005  -8.0827   1.9884  -4.5048   3.1242
#> 
#> (1,5,.,.) = 
#>   6.8643  1.3219  1.0800 -0.0315 -1.4870
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]