Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -0.8538  3.7414  3.5332 -2.9328  2.1515
#>   2.2632  3.5724 -3.3145 -1.7725  4.8423
#>   4.7347  1.4873 -1.3766  1.3687 -1.6326
#>  -2.3290 -3.1808  9.6094  2.9294 -3.8241
#>   0.4746  6.8709  5.9462  9.8163 -0.9700
#> 
#> (1,2,.,.) = 
#>   2.1581 -2.9982  0.3221  1.9249 -1.0838
#>   1.3087  6.8211 -2.2007 -2.4093  0.2833
#>   9.5148  1.7559 -0.6216 -1.7809  1.0024
#>   3.7436 -5.1147  7.2877  4.5028 -1.1264
#>   0.3746  6.7303  1.1461 -2.2281 -0.3031
#> 
#> (1,3,.,.) = 
#>    1.5281  -3.3700  -3.6244   3.3441   3.6080
#>   -7.1669   5.8868   4.8094  -1.5218  -0.6287
#>    1.7669   4.1594   6.9190 -14.0507  -2.9022
#>   10.3544  -7.1211  -0.5756   3.1895   1.5517
#>   -0.4052  -3.7498   8.7138  -2.0279  -1.1154
#> 
#> (1,4,.,.) = 
#>   0.1905  0.3571 -6.2082  2.9589 -1.0372
#>  -3.7305  7.3769  3.5422  2.0008 -4.2623
#>  -5.6070 -2.5181 -1.0765  0.0161  0.1126
#>   2.8082  6.4789 -3.6215  2.6241  2.1958
#>   1.5325 -6.5701  0.0287 -2.1016  1.8735
#> 
#> (1,5,.,.) = 
#>  -4.0822  1.9857  2.3317 -2.7236 -0.4473
#>  -0.1901  5.3316  0.3288 -3.4030 -4.8119
#>   3.6620 -2.7736 -1.4003  2.9507 -0.4975
#>   4.4438  1.5119 -1.9685 -1.7508 -4.0314
#>   0.9647 -0.8674 -5.0017 -3.8060 -5.8403
#> 
#> (1,6,.,.) = 
#>  -3.8189  2.3054  1.4199  5.4591 -1.0336
#>   3.3250  1.9516 -7.8348 -7.3111 -0.7413
#>  -3.0961 -4.7019  9.1205  9.3715 -3.2874
#>  -0.0749  2.1643 -2.7975 -0.7393  3.2122
#>   1.0401 -4.0131  0.0216 -3.7277 -1.9028
#> 
#> (1,7,.,.) = 
#>   -0.7516  -5.6381   3.7772   2.2334   1.9075
#>   -3.9085   7.4768  -1.8531   1.4972  -1.7912
#>    8.6908  -9.5525 -10.1153   5.9635   4.0037
#>    3.1381  -2.4838  11.4102   2.5205  -4.8662
#>   -0.8622  -0.9530   0.3595  -3.5274   0.1391
#> 
#> (1,8,.,.) = 
#>   -0.1224  -4.5458  11.8157   5.7696  -2.7919
#>    8.9622  -2.4798 -11.6338  -1.5878   6.4664
#>    4.1569   1.0449   1.9667   0.7428  -2.1160
#>    6.9810  13.2036  -6.6915  -3.6760  -0.6329
#>   -0.1250   1.0187  11.3486   0.7393  -5.7996
#> [ CPUFloatType{1,8,5,5} ]