Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.3004   6.8511 -13.9843   8.8665  -1.2112
#>    2.6993  -6.0390  -1.0064   2.5759  -5.8028
#>   -3.6901  -2.4639   8.9536   2.8015  -2.3019
#>   -4.6343 -13.0392  -7.5457  11.5272  -2.5308
#>    6.5969   0.4792  -9.8917   4.5190 -11.9744
#> 
#> (1,2,.,.) = 
#>  -0.7953 -1.0338  9.9686  3.7310  2.7262
#>  -5.3577 -4.9447 -0.0474 -2.9904 -0.3750
#>   9.7428 -0.2815  3.2608 -7.2350 -2.5244
#>   1.0668  4.0184  6.3928 -2.3891  2.1963
#>   4.2377  0.9493  4.8657 -5.2937  5.1647
#> 
#> (1,3,.,.) = 
#>  -1.6596  7.4792 -1.1082 -3.3666  8.3487
#>  -11.8616  2.7498 -1.9954 -9.2718 -0.8220
#>   6.8503  2.8988 -1.4653 -0.0235 -5.3887
#>  -7.7606  5.2483 -10.3956  5.5821  8.3730
#>  -3.3476 -2.0096 -2.8800 -3.6141  7.7893
#> 
#> (1,4,.,.) = 
#>  -4.9024  3.7925 -3.7252 -3.6548  1.9346
#>   5.5507 -2.8620  7.5967 -5.6497 -0.4274
#>  -0.0923  1.8599 -0.2282  2.4580  0.4156
#>  -9.3249 -11.3977 -11.5559  3.8884  4.8984
#>  -0.5631  1.0389 -2.6806 -3.9047  4.0335
#> 
#> (1,5,.,.) = 
#>   4.7417  5.9077 -5.5170  1.1475  1.8539
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]