Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -1.5463  0.8859 -2.4614 -0.6957  1.0338
#>   7.8512 -0.1954  3.9685  0.5876  0.0059
#>   0.3233  2.0976 -4.0334  0.3869 -0.2666
#>  -1.1057  1.9271 -0.1180 -4.6673 -2.2886
#>   1.0993  4.8372  0.5181 -1.8210  0.7569
#> 
#> (1,2,.,.) = 
#>   0.4794 -2.8376 -3.3679 -3.1032  3.6640
#>   4.9155  1.4815  2.6609 -1.3137  0.1427
#>   0.3323  2.3293  0.4163 -2.9260 -2.6981
#>  -2.0631  0.6170  2.7899 -5.4704 -5.4411
#>  -0.9742  3.4945  7.6704  4.1578  0.4790
#> 
#> (1,3,.,.) = 
#>   -0.5861  -4.2011  -6.1841   0.0767   4.8489
#>   -9.6325   2.9058  10.0678   5.3364   2.3781
#>   -1.9690   5.6934  -4.2102  -0.0585   2.3930
#>    0.4966  -5.6841   4.5289   4.2648   2.1706
#>   -4.6059  -7.9859  -2.0000  -6.4429  -7.2989
#> 
#> (1,4,.,.) = 
#>   8.7895  2.0862  2.0409  7.5086 -4.4102
#>  -12.2170  0.9026 -5.0235 -2.6899  0.7730
#>  -3.5566 -1.2752 -3.9081 -0.4706 -10.3306
#>  -1.4686 -4.5276  4.4566 -9.9412  5.5643
#>   1.7445 -1.3669 -0.5191  8.4664 -2.7638
#> 
#> (1,5,.,.) = 
#>   0.9550 -2.6123  0.4057  5.9900 -2.1556
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]