Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   6.2238 -3.9629  4.3890  8.8720  6.6126
#>  -1.6234 -4.1307  5.3745  6.3125  0.8285
#>  -8.3057  0.4886 -7.2864 -9.4156  5.4007
#>   8.3346 -1.2873  3.2826 -3.0915  2.5149
#>  -5.5872 -1.0402  3.4008  2.3031  2.7047
#> 
#> (1,2,.,.) = 
#>    1.0833  -8.4520  -6.1782   1.2048  -7.7993
#>   -0.4408  -9.0434  -4.5444  10.6181   0.1840
#>   11.9858   6.4303  -0.9937   4.7320   1.6724
#>   -4.4011  -5.6281  -9.3395   4.0443   0.2006
#>   -0.3313   0.9176   0.0123   2.6495   1.0452
#> 
#> (1,3,.,.) = 
#>    0.0935   9.3062   5.5299  -0.0464   3.2090
#>   -4.7382  -2.6955  -1.1942  -7.4316  -5.5466
#>    1.1181   2.0759  14.6243  -0.4581  -5.0764
#>    2.0662  -8.9279  -2.8356  -0.4320   7.1629
#>   -5.8763   1.3851   0.0220  -5.1136  -1.9992
#> 
#> (1,4,.,.) = 
#>  -1.7158  4.3087  4.3220  1.5615 -2.3465
#>   4.5706  4.3334 -3.3241  3.8890  0.7256
#>   0.2643  2.3448 -5.0431  6.1387 -0.8246
#>  -8.0563  7.4084 -1.0472 -5.0067  1.8074
#>  -2.6589 -2.0440  6.6955 -6.6269 -5.0906
#> 
#> (1,5,.,.) = 
#>   2.0646 -0.7137 -3.0396  7.5220  1.1670
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]