Conv_transpose2d

Conv_transpose2d

Arguments

input
NA input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
NA filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
NA optional bias of shape \((\mbox{out\_channels})\). Default: None

stride
NA the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
NA dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
NA additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
NA split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
NA the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=None, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See ~torch.nn.ConvTranspose2d for details and output shape.

.. include:: cudnn_deterministic.rst

Examples

# \dontrun{

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)

#> torch_tensor 
#> (1,1,.,.) = 
#>   3.1282  4.7949 -5.5622 -1.0866  3.7899
#>   3.4914  4.8807  4.3209  0.8437  7.1147
#>  -5.0276  7.5935  2.4507  6.2129  4.9112
#>   2.9302 -4.0742  1.0907  0.0252  4.4256
#>   4.1733  6.6849 -0.1333  0.7716  1.2488
#> 
#> (1,2,.,.) = 
#>    4.0475   8.1682   9.4413  -3.4628  -2.3695
#>   -3.0556  -4.0963   5.6845   2.0032   0.2438
#>    1.7169  -2.3353  -2.5287  -5.3750  -4.0894
#>    1.7329  17.4464  -1.9850  -1.2224  -1.0126
#>   -8.3888   0.5081  -5.4379  -7.7908   1.4902
#> 
#> (1,3,.,.) = 
#>  -4.6889  1.3331 -3.8890 -4.2812 -1.6408
#>  -5.7474  6.4888 -0.3864 -0.5556  3.3423
#>  -0.5830 -4.7014  0.4339 -4.4822 -0.9338
#>  -3.2573 -5.3475 -6.7339 -4.1705  5.4993
#>   1.4175  5.6303 -1.1562  5.8984  3.9368
#> 
#> (1,4,.,.) = 
#>   2.7974 -1.8220  1.8960 -2.4363  9.3931
#>  -0.0791  9.0332  2.4753  6.5632 -1.9094
#>  -0.4198 -4.7226  4.5077 -6.0814  0.9503
#>  -0.6672  3.4472 -9.0451  1.0115 -4.7566
#>   6.7951  6.0656  9.2166  3.3023  1.2087
#> 
#> (1,5,.,.) = 
#>   4.9513 -5.1344  0.4485  2.9806  0.6510
#>   2.6860 -0.6071  5.0654  6.0352 -0.5143
#>   3.0599  2.8382 -1.2406 -3.1389 -6.3846
#>  -5.3770  3.7280 -12.7695 -4.8459  3.4087
#>   1.6425 -3.5262 -2.5308  5.2363  0.8194
#> 
#> (1,6,.,.) = 
#>    1.6877  11.4338   1.1768  -0.3375  -1.5256
#>   -3.2510  -1.9791   0.4848 -15.0722   2.7618
#>    0.3499   4.9010  -0.0095  -6.5474  -4.1558
#>    3.7509   8.8759   8.5394   2.6775  -2.9372
#>   -6.3247   2.4670  -8.6403  -6.7153  -1.6774
#> 
#> (1,7,.,.) = 
#>   4.5894  2.9587  3.1108  0.5861 -1.6432
#>   0.2160 -2.0561  2.0616  0.3782  0.4569
#>  -0.6198 -5.6256 -4.3481 -0.5680 -6.8947
#>  -1.1064  3.6694  3.1163 -0.8316  5.3151
#>  -2.6586 -10.2829 -6.7049  1.4970  0.0378
#> 
#> (1,8,.,.) = 
#>  -0.6657 -4.0047 -1.7929 -2.8312 -5.8341
#>   0.8324  4.9875 -2.9101  4.2271 -0.9717
#>  -2.1073  2.4393  0.2561 -7.6416 -2.0861
#>  -0.1300  5.5534 -0.9693  0.1627 -0.2768
#>   6.7889 -0.3533  3.2501  3.3338 -0.3276
#> [ CPUFloatType{1,8,5,5} ]

# }