Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -1.9338 -3.7021  8.5338 -4.4989  2.7826
#>  -7.4790  0.2098  5.3342 -3.1822  1.4282
#>  -10.8049 -8.5130 -4.9559  1.5548  2.6806
#>   0.3373  4.9267  4.5645  5.7335  5.2444
#>  -3.3644 -11.0871 -3.8517 -3.8187  3.4132
#> 
#> (1,2,.,.) = 
#>    5.0560  -7.5617   2.3165  -4.8879   1.4759
#>    5.2124   1.4226  -0.1040  -7.3305  -3.6062
#>   -0.2930  15.1274   5.5031  -7.8129  -0.9362
#>    7.5847   9.6668  -2.5559   2.8369  -1.0545
#>   -1.1707  10.5578   3.6621  -4.0282   3.0225
#> 
#> (1,3,.,.) = 
#>  -1.7931 -5.9532 -6.5420 -5.4370 -0.7775
#>  -1.0745  2.6567  0.1473  9.2200  1.0498
#>  -0.5491 -0.1064 -3.3667 -3.3579 -7.5647
#>   3.4197  1.9489 -0.6255  1.2519 -2.2706
#>   1.1321  4.5615  1.9728 -0.2816  2.1347
#> 
#> (1,4,.,.) = 
#>   -4.1777   1.3827   1.4811   2.6734  -0.9167
#>   -5.7129  -9.8017 -10.1762  -9.6407  -2.9062
#>    5.0887  12.0977   2.7699   3.0581   4.3126
#>    2.9646  -0.7131  -6.4739  -3.5694   2.3212
#>   12.9849   7.2898  -2.0740  -5.6767  -5.4776
#> 
#> (1,5,.,.) = 
#>   0.9808 -3.9458 -3.5571 -5.5565 -3.4089
#>  -1.5229  4.5125 -2.0509  3.2504  3.4137
#>   4.9765  5.1691 -6.2621 -8.4066  2.5630
#>   3.7904  3.7125  3.7217  4.5440  2.7167
#>  -1.9355  1.2713 -8.0517 -2.6923  1.4007
#> 
#> (1,6,.,.) = 
#>  -5.6988  1.9119  1.4541  1.0408 -3.1010
#>  -3.8535  6.4702 -3.9750 -14.8557 -0.6974
#>  -0.0934  3.4193 -6.1292  3.3779  5.1750
#>   2.1070  6.3509 -7.7881 -5.8317  1.4205
#>   2.6407 -0.3862 -2.2048  3.0411  0.2656
#> 
#> (1,7,.,.) = 
#>  -0.6518  4.6987  2.6808 -4.0481  1.9288
#>   8.3000 -0.3874 -0.2392  6.5368  8.3624
#>  -0.4931  1.1724  3.3669 -0.8063 -1.6890
#>   4.9329 -4.1507 -3.8371  3.3669 -0.9805
#>   5.0687 -2.0841  4.5801 -2.5867 -1.6759
#> 
#> (1,8,.,.) = 
#>   -0.5794  -4.2049  -6.1545  -2.9992  -7.9937
#>   -1.0575   5.9393  -2.3519  -4.2976   1.0864
#>    0.3222   5.2918   3.6195   5.2018   3.1489
#>   -2.5679  10.4189  -3.5122  -8.6371  -2.2733
#>   -2.6154   2.8460  -0.1623   5.9856   1.5642
#> [ CPUFloatType{1,8,5,5} ]