Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -0.5949 -0.0351 -1.4653 -0.9360  7.4637
#>   0.4176 -4.7256 -3.5665  0.5890  5.5792
#>  -0.8308 -2.1385  1.9010  0.8598  4.2204
#>   2.4519  4.7937  1.4380  3.8313  5.8333
#>   6.7353 -0.5539 -5.1339  4.4604 -2.4983
#> 
#> (1,2,.,.) = 
#>  -0.5859  2.6933  2.7077 -1.4659  2.5188
#>  -2.0084 -1.9321 -0.0199 -0.5801  2.6232
#>   0.7850  1.8131  4.9003  3.5108 -10.7643
#>  -1.1783 -5.9186 -1.1071 -2.1876  1.2400
#>  -1.8715 -1.0213  0.4519 -4.0206  2.5528
#> 
#> (1,3,.,.) = 
#>  -2.1988  0.1903  2.7555 -8.5033 -0.5145
#>   1.3725  1.7354 -7.8205 -4.7546  4.8916
#>  -8.5988 -7.7405 -8.4673  8.7754 -1.8867
#>   2.2745 -2.4401  4.9849 -0.6849  6.1009
#>   4.1183  4.7554 -8.2425  9.3508 -2.4751
#> 
#> (1,4,.,.) = 
#>   -1.1504   1.8600  -2.1236  -4.0108  11.6244
#>   -1.9880  -2.9878   0.4903  -2.3271  -4.4168
#>    0.4877   6.5101   0.9080   0.2264   1.3348
#>    9.6944   5.2813  -8.5883  -1.3569  -5.2216
#>   -4.5563  -0.7308  -1.4536   0.4295  -7.0908
#> 
#> (1,5,.,.) = 
#>   0.1062  1.1234  3.8977  3.1539  4.5729
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]