Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   3.5693 -1.1108 -4.5684 -6.8499 -1.1365
#>  -1.1156 -6.5258 -7.6609  4.1570  1.4219
#>  -2.9595 -1.7642 -0.5510 -8.9429 -5.2902
#>   3.1903 -1.4573 -5.2864  8.5647 -11.9970
#>   3.2928 -8.1826  1.8711 -4.5162 -7.3973
#> 
#> (1,2,.,.) = 
#>   -5.2377   6.7191  -5.5583   0.8232 -12.2707
#>    3.8005  -0.7898   2.9730   0.4118   2.1843
#>   -7.2949  -7.2575  -4.0482  -1.5095   6.9239
#>  -13.9025  13.3171  -1.1829  -7.7236   3.3027
#>    4.0450   9.0058 -11.1968   1.3764   5.2272
#> 
#> (1,3,.,.) = 
#>   1.1988 -0.7666  3.1428 -3.7930  5.6576
#>   0.1850 -1.0045  4.8356  2.8577 -9.1682
#>  -0.8810 -3.8922 -4.1319 -3.9258 -2.3467
#>   2.4025 -11.9842  1.7019  7.0931 -1.0340
#>   3.3151  4.5727 -2.7969 -2.4867  3.9587
#> 
#> (1,4,.,.) = 
#>    3.2374  -6.0574   3.3111   3.2034   1.7647
#>   -1.6022   5.1908  -2.0604  -6.8458   0.6835
#>   -1.2655   4.3783   0.7415   5.9552   6.7290
#>   -2.7275  -6.9396   0.6561   2.3907   0.0401
#>   -5.3824  10.7670   4.5976  -2.8406   6.0287
#> 
#> (1,5,.,.) = 
#>  -3.9844  5.6423  2.9532  2.0186  5.7272
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]