Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -3.2440  1.2664 -8.9611  1.0623  6.7182
#>  -0.2448 -4.0049  1.3709 -0.2084 -0.2630
#>   1.1464 -0.7051  0.9806  2.9289  1.3516
#>  -2.6273  5.5504 -4.3216  4.6602  4.1616
#>   1.4053 -6.3723  0.6322 -2.7466 -2.2938
#> 
#> (1,2,.,.) = 
#>  -0.5303 -5.4287 -1.0481 -4.1302  2.1112
#>  -3.9566 -2.7993  0.9098 -1.9640  0.9519
#>   8.2682  1.9613 -0.6034  9.1754  2.7704
#>   2.4369  4.7060  1.1603 -4.9755  3.4807
#>   5.3864 -4.5563  0.0924  0.1186 -4.1471
#> 
#> (1,3,.,.) = 
#>  -1.7737 -2.3338 -5.4699  2.4637 -2.0549
#>   6.8782 -7.0181 -0.1494 -8.2514 -4.3902
#>  -3.4681  5.2279 -3.0125 -13.2124  8.7259
#>   2.6481 -5.7475  8.3335 -2.3802  4.2669
#>   0.1484  1.2837 -0.0800  0.2536  2.0959
#> 
#> (1,4,.,.) = 
#>  -1.1879 -5.6196  4.2347  3.0754 -1.3636
#>   7.1301  8.4678 -9.0858 -9.2797  4.1535
#>  -3.5998  4.8522  4.0590  8.0088  8.9454
#>  -2.8425 -2.3888  6.0017 -13.8310 -6.3949
#>   0.0687 -3.4743 -0.7834  6.1250 -1.0978
#> 
#> (1,5,.,.) = 
#>   3.8388 -2.4738  0.7748 -8.1330 -3.2729
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]