Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -4.8727  -2.6136   3.1428  -7.0745   1.1706
#>   -7.6407   8.1840   2.8487  -6.4339   3.9525
#>    1.8357  -6.0645  -3.4667  -2.9826   1.3245
#>    1.1596 -16.6961  15.7248  -3.6209   6.4159
#>    4.4550  -5.1847   7.0886  -2.6652   5.0840
#> 
#> (1,2,.,.) = 
#>   0.9503  0.5763 -6.0498  2.5525 -3.5487
#>  -3.1032  5.8947 -11.8099  9.8238 -0.5601
#>  -4.4131 -9.4170 -5.2627 -1.3059 -0.9175
#>   1.4627  2.4417  1.4949 -0.0702  0.7204
#>   0.2459 -2.9828 -1.5780 -2.6176 -2.4656
#> 
#> (1,3,.,.) = 
#>    0.6114   5.1240  -1.6108   0.1415  -2.6727
#>   -4.2525  11.9059  -5.9150   9.4957  -4.6880
#>   -2.9150  -6.3546  -0.3568  -3.4386   3.2000
#>    7.6725 -21.8564  24.5866 -16.2865   8.8586
#>    8.3547   2.7167   3.7878  -9.2188  -5.6644
#> 
#> (1,4,.,.) = 
#>    1.2608   0.2468   0.3269   2.4460  -1.1105
#>   -0.7610   5.6234  -4.0955   2.6785 -10.0584
#>    5.4450  -2.9893  -2.2623   0.9415   1.0945
#>   10.7976 -17.6679  10.8951  -8.0319  -6.9381
#>    2.2222 -10.6992   2.0539  -1.9295   0.4741
#> 
#> (1,5,.,.) = 
#>   0.9262  3.3788 -1.3763  3.5654 -4.7522
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]