Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -5.0711  8.7516 -9.9199 -5.1308 -2.1660
#>   7.6302  0.8472  2.3739  2.5454  1.6630
#>  -6.1582  7.2304  6.3747  4.8602 -7.3775
#>  -1.5190  0.9856 -7.7851  8.0220  6.5719
#>   0.4236  4.8358 -1.9878 -4.0798 -4.1261
#> 
#> (1,2,.,.) = 
#>   -6.8751   3.1362  -3.3014   3.5066   1.6799
#>    6.0476  -3.8349  -0.2652  -8.6113   3.9531
#>   -3.0692   3.1299  -9.2314   8.4970  -2.2383
#>   -0.8714   4.7314  -0.3461 -13.9887   5.9202
#>   -2.5805   0.5041  10.1957  -2.4107  -1.9888
#> 
#> (1,3,.,.) = 
#>   6.0416 -2.8504  6.3344  4.8618  4.2778
#>  -2.7819 -4.9073 -9.5587  9.3698 -0.5376
#>   1.6583 -0.4790 -2.9546 -0.1291 -1.2382
#>  -1.3385  3.4277 -0.8989  2.1624  1.2546
#>  -0.0547  3.4967 -2.9231  4.0580  4.9724
#> 
#> (1,4,.,.) = 
#>   2.7673 -3.5999  5.2915 -4.1294 -6.8392
#>  -5.3250 -1.9549  3.1452  4.0069 -4.7296
#>   2.1425 -1.5783  1.0085  0.9109  3.8709
#>  -5.3542  2.0018 -4.0449 -4.1816 -8.6263
#>  -0.8711  4.8077  4.1169 -3.6909 -2.4701
#> 
#> (1,5,.,.) = 
#>    0.4842   0.2211   5.3512  13.1898  -0.8758
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]