Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -13.8169  -4.1826   3.1586  14.9793   6.8078
#>   -1.5083  -2.0213  11.9943   4.8958   6.7112
#>   -0.4414   0.5513  -4.1075  -4.7333  -2.2741
#>    0.4874  -3.8230   5.8018   4.0625   0.9874
#>   -2.3512  -0.5123   0.8273   1.7563  -0.1801
#> 
#> (1,2,.,.) = 
#>    3.5830  -6.4038   1.6304  -1.3564  -3.4466
#>    6.3028  16.3055   3.4315  -2.3199 -16.1657
#>   -7.0179   2.7302   9.8803   2.3266   3.3474
#>   -2.1244   2.2193  -0.5536  -5.7889   0.3057
#>    3.7474   5.7272   2.3822  -7.3542  -0.8449
#> 
#> (1,3,.,.) = 
#>    2.2985  -9.1995  -1.8768  -9.0964   0.2614
#>    3.9264  15.0722   8.0639   7.5346  -3.6081
#>    5.6210   9.9884   1.1232  -4.4402  -3.5815
#>    0.4411  -3.0348   4.5728   1.3823   4.1485
#>    4.0763   2.5107   2.2054  -5.4110  -0.6208
#> 
#> (1,4,.,.) = 
#>  -3.6181 -2.7926  4.8356  0.7424 -0.8744
#>  -0.3663  7.2354  0.2006  4.2472  2.3453
#>  -6.9217  4.1762 -5.8611 -4.3888 -2.1666
#>   2.6334  3.3437  9.5233  4.8604  3.2165
#>  -2.5106  0.1039  0.2328 -0.8685  0.6957
#> 
#> (1,5,.,.) = 
#>   -5.6300  10.3104  -4.7099   0.9242  -6.6001
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]