Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   3.1260 -3.4989 -8.5065  9.1216  4.1147
#>  -2.2244 -14.3416  4.9415  2.1154 -1.0645
#>   2.5649 -5.2622 -4.7593 -5.8941  6.2561
#>  -0.9735  7.7534  8.2427  7.8354 -0.7387
#>   3.8614  2.6768 -0.0831  0.8909 -3.2692
#> 
#> (1,2,.,.) = 
#>   -6.6590  -7.4497   7.4599  -4.9450  -8.4245
#>   -1.2084   3.6497   2.9882   4.6876  10.7132
#>   -5.1365   3.0339  11.1175  -2.3023  -1.3611
#>    4.6781  -7.3703 -14.4831  -7.6848  -9.3073
#>   -5.9746  -1.0039   6.2093  12.4989  -5.1423
#> 
#> (1,3,.,.) = 
#>  -2.1704 -0.3456  6.8894  1.8100 -5.6366
#>  -0.0979  4.1192  3.1770  4.9780 -8.0628
#>   2.5511 -8.7862 -4.7620  2.5104 -0.1565
#>  -2.8573  8.1476  3.5438  4.0888  1.4817
#>   5.7448  6.8273  1.6640  2.2603  3.0117
#> 
#> (1,4,.,.) = 
#>   -0.8304   4.0343  -1.0077  -6.5408  -0.4150
#>    2.4334   4.5415   4.3144   3.7588  -4.3927
#>    0.5422   1.6788   0.1564  -1.5540   5.9669
#>   -1.3468  -1.5543   2.3431   0.2763   3.0825
#>    2.0719  12.0093   5.6280  -1.9382  -6.2889
#> 
#> (1,5,.,.) = 
#>  -0.6824  7.0742  1.7148 -10.5340 -1.1377
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]