Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    1.1320  -4.9633  -6.6041   1.8266  -2.0651
#>   -2.8150  -4.0796  -2.6084  -4.4087  -5.0851
#>   -9.9182   1.0302  -7.9658   3.2908   1.0181
#>   -2.8732  -4.1595  13.0834  -3.3304  -2.0692
#>    2.1888   2.0896   0.4990  -1.2823   2.3918
#> 
#> (1,2,.,.) = 
#>   1.1402  0.0300  2.4739 -7.7193 -2.7781
#>  -1.6955  3.1455 -0.8854  1.1782 -6.8796
#>  -5.9864 -3.3368 -0.7326 -4.2648 -1.7477
#>   0.6620 -1.5435  4.7606  0.3075 -7.6350
#>   6.2701 -5.9817 -2.5457  3.6343 -7.7500
#> 
#> (1,3,.,.) = 
#>  -4.8724  3.2676 -0.7124  6.7249 -3.1708
#>  -2.6456  6.2512 -4.6399  3.6060 -0.0919
#>  -0.6719 -5.5602  6.1121  0.6100  0.6432
#>  -0.1520 -1.5409  8.4894 -0.4903 -0.5162
#>  -2.2670 -5.0786  3.1157 -0.6778 -3.7087
#> 
#> (1,4,.,.) = 
#>   0.4388  5.2726  1.2182  2.2016 -0.2535
#>   0.0667 -2.3661  7.5798 -2.3821  6.9956
#>   4.6191  0.3391  2.1481  5.5177 -1.6430
#>   4.2160  0.0510 -1.2307  5.7143 -2.7585
#>  -1.5289  1.1373  0.2714 -2.3286 -1.3905
#> 
#> (1,5,.,.) = 
#>   5.4494 -0.2388  5.8024  9.8208 -0.9314
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]