Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -1.8835  0.3325 -1.7097 -1.8178  5.8665
#>  -5.2408  1.0799 -5.1443  1.4088 -1.8045
#>  -1.1190 -4.8990 -0.8335 -6.3993 -3.3608
#>   0.3697 -0.3993  2.7437 -1.8654 -3.5979
#>   4.3747 -1.8038 -2.4185  2.6496  1.3140
#> 
#> (1,2,.,.) = 
#>  -1.4566 -3.7693 -2.5759 -0.0338 -9.2811
#>   5.2632 -9.8000 -1.7590 -4.9831  3.9439
#>   1.1036 -2.8600 -2.3730  1.3056 -4.8800
#>   4.3990  7.0599  9.0623  4.8147 -3.2450
#>  -4.5880  1.1231  7.4219 -3.5181 -0.0203
#> 
#> (1,3,.,.) = 
#>    1.9651  -1.2305   4.6358  -0.4601   0.9546
#>    0.4872  -0.7996   3.4670   8.3581   2.4513
#>    0.2410  -2.1106  12.5688   2.1233  -7.3299
#>   -1.6222  -1.8719  -9.5484  -4.5681   0.6272
#>   -0.5065   4.0345  -5.6359  -0.8137  -1.1686
#> 
#> (1,4,.,.) = 
#>  -2.1845 -4.0721 -3.8660 -9.1237 -2.4158
#>  -3.8758 -5.6345 -9.8500 -3.4471  1.9237
#>   5.9653 -0.7038 -5.7221 -5.8351  3.0693
#>  -4.7049 -0.0911  1.3265  2.3894  1.3786
#>   4.2743  6.1147  6.9928  4.3687 -4.0049
#> 
#> (1,5,.,.) = 
#>   1.0512  4.0228  3.8316  0.5194  2.7077
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]