Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    3.2715   4.8964   0.4631   0.4154  -2.6138
#>   -9.0496   0.3911   1.1300   3.6364   4.3387
#>    8.4117  -4.5971  11.3322   4.6914  -0.4124
#>   -1.3834  -3.7734   5.4164  -3.4501   0.3331
#>   -3.7031  -6.1763  12.9816  -5.0137   1.9518
#> 
#> (1,2,.,.) = 
#>    4.3813  -7.4388   6.7569  -2.9604  -2.3838
#>    4.5567  12.3194   1.0736  -2.2936  -5.7118
#>   -7.1203  -7.8273  -4.8365  -0.5515   6.9255
#>    8.5654   0.4702   0.0805  -8.3459  -2.2066
#>  -10.3966  -7.0789   4.5950  -0.8606  -4.2465
#> 
#> (1,3,.,.) = 
#>   -3.7760  11.4161  -4.8356   2.4160  -4.8002
#>   -3.4529   8.7056   6.6939   1.9042   2.6077
#>    5.5488   8.4479  -2.2897   5.4284   8.2467
#>   -6.7756  -2.4236  -7.8646  -2.4106   1.2853
#>    5.8532  -4.7417   2.2488   3.1966  -5.0621
#> 
#> (1,4,.,.) = 
#>    8.8840  -4.6097   2.7095  -1.5452  -4.6142
#>    1.6255  -5.9398   0.0967  -7.0369  -3.4573
#>   11.6370  10.0142  -2.2256  -2.7223   5.9886
#>   -7.1956  -0.2988  -2.2416   6.1061   9.7389
#>    3.0529  -0.5178  -2.5679  -5.9032   0.4686
#> 
#> (1,5,.,.) = 
#>   10.6783  -6.0734   2.0918  -4.2169   3.3602
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]