Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   2.9126  1.8106  3.2510 -6.3951 -3.7319
#>   2.7582  8.9874 -8.0080 -4.3576 -9.1519
#>  -7.9617  2.3853 -5.5400  3.7663 -0.3554
#>   7.1510 -2.7818 -3.0964 -0.8162  0.9934
#>   2.7403  0.7240  1.3078  4.5795 -1.7954
#> 
#> (1,2,.,.) = 
#>    6.0364   7.3679   8.4879  -3.7754  -2.7436
#>    3.0347   5.0615  -3.1218   3.9014   3.4821
#>   -2.8457  -1.9765  -7.9165   0.0699   4.6415
#>   12.5752  -1.0754 -15.0638  -0.2671   0.9033
#>   -7.2930  -5.9320   9.3342  -1.9257  -6.9981
#> 
#> (1,3,.,.) = 
#>   -5.4445   5.7359  -0.5881  -4.3803  -9.1979
#>    3.5341  10.0550  14.5759   3.1902  -6.5696
#>  -10.5693 -12.8649  -1.8934   1.2406   9.1742
#>    3.0285   5.1074  -6.6789 -11.7873  -3.2232
#>   -1.2519   8.3080   4.4600  -2.4729  -1.6967
#> 
#> (1,4,.,.) = 
#>  -4.9064  5.8141  6.1061  3.8643  4.7319
#>   2.7308 -4.2439 -5.8656 -2.9043  0.6089
#>   0.1323 -1.4049 -3.2863 -12.4736 -3.6276
#>   0.5456  4.5971  0.7013  3.5413  2.3783
#>   1.3441  0.5689 -4.0116  1.5774  6.7455
#> 
#> (1,5,.,.) = 
#>    2.8767  -0.4817  -5.3908   1.0965  -6.2579
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]