Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -4.3913  3.3129  0.1794 -7.3480  2.4302
#>   0.3734  3.5978  0.5481 -2.5312  8.4165
#>  -5.8181 -0.0178 -4.9177 -6.3276  0.3138
#>  -5.3807 -5.5892  8.8366  8.9832  4.1677
#>  -2.6134  7.4516  3.5584  2.1115  2.4546
#> 
#> (1,2,.,.) = 
#>   8.2627  2.0691  0.0903 -3.5413  1.1877
#>   0.7264  2.0448 -8.4740 -1.2162  0.3867
#>   1.1053  0.7159 -1.4128 -0.7986  0.8012
#>  -3.5495 -1.7035 -0.9313 -1.5449  4.5175
#>  -6.4891  3.4214  1.4274 -1.9420 -0.8753
#> 
#> (1,3,.,.) = 
#>   -4.1923   1.3053   2.7614  -2.4275  -0.7393
#>   -3.8340  -7.8581   4.5604   3.0962   1.0068
#>    4.6379   1.5661  -0.7784  -1.1726  -2.7959
#>    1.2678 -10.8736  10.0538   4.1340  -3.7063
#>   11.3219  -3.3171  -5.5358   0.0823   3.1648
#> 
#> (1,4,.,.) = 
#>   -2.6694   1.1179  -0.3586  -0.4494  -4.3537
#>    0.0424  -2.7736  11.3824   0.3703   4.5371
#>    2.0110   1.7219   0.1291  -1.6645  -8.3837
#>    6.0557  -1.6544   6.4642   4.2688  -4.7350
#>    3.4928   3.4208  -0.0798  -2.1485  -1.1263
#> 
#> (1,5,.,.) = 
#>    6.1617  11.4112   0.5538  -4.7722  -2.9251
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]