Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   6.2831 -1.8390 -1.5741 -2.8062  3.3178
#>  -14.6471  8.8640 -7.0061  0.6117 -3.6830
#>   5.1926 -9.9984  7.8774  3.0359 -5.6046
#>   4.3478  3.9136  0.2813  6.7857 -3.6405
#>  -3.0773 -1.0123 -8.2707 -4.1931  2.0319
#> 
#> (1,2,.,.) = 
#>   1.6830  5.7390  3.9952  5.8474 -1.0152
#>   1.9510  8.3528  1.9273  1.7293 -2.2963
#>   2.8716 -5.0377 -0.1472 -4.9129 -1.3247
#>  -1.4301 -0.1503  0.9646  3.7979  5.2279
#>  -2.6117 -9.3358 -2.1337 -5.6376 -0.8977
#> 
#> (1,3,.,.) = 
#>   -2.9324   8.5216  -1.8526   3.9822  -2.3510
#>    5.5657   0.4664  -6.2818   2.5147  -7.0319
#>    2.3646  -8.0812   7.3683  11.1403 -10.7945
#>   -1.0423  -2.2627   1.5491   6.1278   4.7604
#>    1.4737   0.4208  -4.2947  -2.6900   6.5342
#> 
#> (1,4,.,.) = 
#>  -2.7060 -6.2994 -4.4504  0.6062 -0.6091
#>   5.0105  3.1979 -6.9195  3.4038  0.0900
#>  -1.1699 -1.2935  4.8001  4.1651  1.0828
#>  -0.4051 -6.8208  0.4415 -1.6388 -4.0021
#>   2.9851 -4.8291 -3.5050 -0.3336  2.4253
#> 
#> (1,5,.,.) = 
#>  -0.1967  0.6294  6.3842 -2.1862 -1.6914
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]