Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   -1.4123  -1.4836  -5.6028   7.3692  -2.7814
#>   -5.9963  -2.6366 -12.4501   1.7221   0.9943
#>   -0.8824   0.6426  -8.5220   0.6371  -2.0738
#>   -5.1027  -6.0810  -1.4834  11.3255   0.7765
#>   -7.7355   0.9061   5.4948   9.5553   0.9418
#> 
#> (1,2,.,.) = 
#>  -1.0076  2.1188 -1.7260  0.7297 -0.0542
#>   0.9405  0.7754  4.3360  0.6900  2.9745
#>  -4.5046  0.6897 -3.3852  4.9404  7.8055
#>   2.7808 -9.0293 -4.6783 -6.2424  3.1814
#>  -1.2841  0.1817  6.1271  5.7825 -2.5033
#> 
#> (1,3,.,.) = 
#>  -3.1357 -8.2102 -3.9501  1.3307 -5.9977
#>  -1.4810 -5.2979 -20.8132 -7.4497 -10.5755
#>  -3.9926  1.0181 -6.9951 -1.5550 -4.3954
#>  -5.6866 -6.3416 -2.9149  6.0030  3.4344
#>  -4.0958 -9.5107 -3.8027  4.5415  3.3585
#> 
#> (1,4,.,.) = 
#>    0.7733  -0.5980  -1.8976  -1.7372   0.0161
#>    2.6772   6.6007  10.3038  -1.6739   1.9610
#>   -3.2225  -0.5058  11.4375   2.8543   8.0488
#>    1.7957  -4.8982   2.9715  -4.3380   5.3875
#>   -2.4672  -8.2390  -4.0270  -2.2988   0.7584
#> 
#> (1,5,.,.) = 
#>   -2.1893  -2.3593  -2.5441  -2.0514   2.6414
#>    2.8785  -0.1186   0.5548  -1.2181  -3.8304
#>   -0.8212  -0.8711   8.1637  13.8240   2.5953
#>    1.2499   0.4279   1.2857   4.7750   4.4601
#>   -1.6496   4.2760  -0.0424   0.1821   1.4217
#> 
#> (1,6,.,.) = 
#>    6.0240   3.2671  10.3538   0.3912  -7.8699
#>    0.4714   3.5858   7.3978   5.2481   0.5388
#>    3.4793  -1.2042  -6.3471  -7.0744  -2.7506
#>   10.2497  -0.6451  -3.7747 -10.6539   0.1107
#>    7.1721  -1.3261  -5.3825  -2.1505   0.9749
#> 
#> (1,7,.,.) = 
#>   -3.0149  -1.1188  -1.4328  -9.4478   1.9073
#>   -0.0990 -12.1381  13.0611 -10.3911   7.8598
#>    5.2112 -13.4249   3.6201  -8.3143  -0.7575
#>   -1.6193  -0.8332  14.2088  -4.0620   7.4604
#>   -3.6777  -4.4320  -3.3744  -1.1851   4.4483
#> 
#> (1,8,.,.) = 
#>  -0.6550 -4.1750 -3.5765 -0.7651  2.1225
#>   0.8425  2.9625  3.5868  2.8971  1.9603
#>  -6.7742 -2.5494  4.7967  4.1736  8.6077
#>   2.5869 -3.2647 -2.5393 -8.8532 -2.5787
#>  -1.9755  1.1859  8.4727 -1.2093  1.2136
#> [ CPUFloatType{1,8,5,5} ]