Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -2.4581  -2.8853   0.0212   4.7095  -8.0322
#>    0.3149  -5.2923  -8.4411   1.3768  -3.9489
#>    0.3150  13.3543   6.3844  -8.4515   5.6086
#>    3.0012  -0.0418   0.3849 -13.2813  -7.5657
#>    4.0989   5.4120   1.6170  -8.8428  -1.8900
#> 
#> (1,2,.,.) = 
#>    1.3236   0.3250   6.4605   6.4402   2.2331
#>    2.5989 -18.4042  -8.5294   1.8539 -13.9614
#>    1.5757   5.7187  17.5049  12.0695  -5.2058
#>    7.9105 -15.4041  -2.1173  -9.7301  -8.9027
#>   -0.9533   5.0139   2.3900   5.2575  -1.3625
#> 
#> (1,3,.,.) = 
#>   -1.6463   0.5680   1.1908   1.6216   0.8345
#>   -1.3480  -1.4544  -0.2434   2.0874  -8.2750
#>    4.1697   1.8294  -6.4702  12.9933   0.7486
#>   -2.0680   2.6098  -0.6300  -9.2936  -5.4522
#>    5.2347  -7.5664  -3.1687   1.8802   4.6927
#> 
#> (1,4,.,.) = 
#>  -4.8102 -6.4365 -1.5497 -6.2208 -5.2421
#>   0.9055 -5.0888  8.1634 -2.5243 -1.7668
#>   3.3866 -3.4152 -3.1890 -0.0695 -6.9363
#>   2.7913  3.7401  1.8530  4.5578 -5.8657
#>  -4.3485 -0.5772 -4.2553  3.5192 -3.2425
#> 
#> (1,5,.,.) = 
#>    4.4910  -4.6366  -0.6649  10.7169  -7.5835
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]