Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -2.6206  4.3106  3.0774 -6.6749 -2.2280
#>   7.8457 -1.5635  0.6528  4.3060 -2.6222
#>  -5.9787  1.6884  6.3915  6.1219 -0.8914
#>   0.7600 -2.2499  9.7042  1.5571 -1.4499
#>  -3.8963  1.6930  7.1608 -1.9007  2.0972
#> 
#> (1,2,.,.) = 
#>   4.5732 -6.3117 -0.4619 -0.2964 -4.2020
#>  -2.4082 -3.8891  5.8191  4.5861  2.1638
#>  -0.2392 -1.8831 -0.5068 -1.2910 -4.9857
#>  -1.6010 -6.0638 -0.1133  2.3269 -0.5159
#>  -1.6077 -0.6605 -2.7616  1.1215  1.2590
#> 
#> (1,3,.,.) = 
#>  -0.7658 -3.2726  0.9111 -2.1166 -1.7076
#>  -1.0007  6.7197 -4.4857 -2.4486  0.9922
#>  -3.1810 -2.1988  1.7317  5.4220 -1.8309
#>   1.8319  4.7836 -3.8651 -5.5612 -5.2917
#>   0.5896 -7.2184 -0.9303  3.5426 -6.3130
#> 
#> (1,4,.,.) = 
#>  -1.5884 -1.6201 -2.3101 -1.4193 -0.3321
#>  -4.4976 -1.4712  0.5054 -1.9875 -0.4993
#>   1.7942 -8.8047 -4.6635 -0.8027 -2.8396
#>  -0.4242 -7.2137 -0.1556 -7.9972 -0.8149
#>  -1.3617 -3.0171 -0.2115 -0.7320 -0.4588
#> 
#> (1,5,.,.) = 
#>  -1.4536 -3.6221  6.1019  4.3207 -3.1546
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]