Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.2287   3.3094   3.4743   9.7768 -11.2155
#>    4.6119  -3.9745  -2.0776  -2.6965   7.6324
#>   -3.3576   5.6565   0.8673   3.3209   0.9052
#>   -1.6333   5.8990  -7.6064   8.3301   4.5317
#>   -2.8191   1.0853  13.9369  -1.0195  -0.0228
#> 
#> (1,2,.,.) = 
#>   0.2811  6.0302  3.8766 -1.5118 -2.3007
#>  -16.6919 -14.1872  2.9698 -7.1464  2.1782
#>   2.0997  6.7727  1.4973  4.5542  2.0073
#>   2.2646 -3.2244  4.4951  2.5730 -3.0692
#>  -5.6957 -0.5530  2.5754 -4.1192 -2.7429
#> 
#> (1,3,.,.) = 
#>  -3.6452 -2.6004 -1.8486 -5.5088  0.6443
#>   0.2363  1.0250  1.2637  1.2478 -0.8696
#>   4.5404 -2.5262  3.3706  3.2811  4.8811
#>   1.4631  0.6383 -3.0246 -1.2366  0.7995
#>  -0.6051 -3.3852  1.6983  1.2872 -5.9887
#> 
#> (1,4,.,.) = 
#>   -2.0523   2.4781  -3.7118   3.1201   3.3373
#>    3.8076  12.4211  12.6142  -0.7973  -2.7809
#>   -7.6533 -13.1896  -1.0605   0.5238   3.3664
#>    4.7701   7.6760  -0.9524   7.1656   0.5227
#>   -0.6362  -9.5032  12.0889  10.1195   2.4994
#> 
#> (1,5,.,.) = 
#>    5.6975   5.7209  -0.2738  -2.1517   1.3446
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]