Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   0.3854  4.9360  4.9625 -4.0224  4.8648
#>   4.3514 -10.0582  1.5610 -1.3080  3.2657
#>  -2.0232  8.6031  2.0688 -2.1710 -2.2343
#>   4.5373 -0.8339 -3.6218 -0.5293  1.7947
#>  -0.3458 -5.3760  1.5919  1.2573  4.4634
#> 
#> (1,2,.,.) = 
#>   0.8173  5.6768  0.4426 -8.6542 -1.9723
#>  -3.9717  0.6599 -6.9812 -1.9961 -2.1247
#>  -4.1326 -2.0900  5.9207  1.8169 -0.0905
#>  -4.0141  2.3184 -12.5316 -5.7160 -3.7079
#>   0.2774 -2.6513 -8.9629  3.9414 -0.2130
#> 
#> (1,3,.,.) = 
#>   -1.3630   1.5184  -2.4862   3.3088  -3.2240
#>   -0.7452   2.4529   0.3613  -2.8935  -4.6709
#>    7.3596 -13.3048  -2.0152 -12.5908  10.2597
#>   -7.0884  -2.3000  -1.3713   6.1168   1.4777
#>    5.8392   9.3791   0.9878   2.4798  -4.5572
#> 
#> (1,4,.,.) = 
#>   -5.5675   6.2234   1.4764   0.5642   0.6584
#>   -5.2493  -4.3197   0.0655   2.4404   3.2520
#>   -3.6249  -8.7502   7.1429  10.3124  10.8260
#>   -1.8302   4.3906   7.5001   8.8110  -5.1930
#>    3.3417   5.2108   5.1780   5.4587  -0.9975
#> 
#> (1,5,.,.) = 
#>   -5.8785   0.4711   7.1384   4.6037  -3.8530
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]