Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -0.3742   1.7571  -1.6475  -5.0138  -3.3465
#>    0.3501   0.2154   3.8881  13.1293   1.6660
#>    3.8226  11.1777   4.5330   4.7664   7.2239
#>    4.1329   3.2826   6.9169  -1.0461   4.3229
#>    0.1762  -0.5826   2.0477   0.6623  -2.5236
#> 
#> (1,2,.,.) = 
#>   4.9183 -8.2394  5.3762 -1.6810 -8.0609
#>   3.4410 -3.2349 -1.7780  0.7134  1.2443
#>  -3.4579 -1.9972 -2.0466 -0.5021 -0.4504
#>  -3.2997  2.4932 -1.8747 -3.7774  4.2482
#>  -7.9202  1.5723 -2.4433 -4.8742  1.7495
#> 
#> (1,3,.,.) = 
#>    5.8191   3.2834  -8.3490  -3.2309   4.2086
#>   -3.2507  -0.3183   3.1731  15.1854   3.3860
#>    9.9391   9.0359  -0.9365   5.0864  -1.1071
#>    4.9968   4.1393   2.3373  -1.0039   3.4839
#>    5.0084   3.3821   6.5704   8.5769   1.7994
#> 
#> (1,4,.,.) = 
#>   2.6676  8.8763 -5.3605  2.0021  0.6022
#>   1.5331 -5.3108  9.3879 -6.2079  7.1915
#>   8.0538  0.1004  2.4127 -0.8095  0.3138
#>   0.3616  6.3834  4.6532 -1.9583  0.8827
#>  -0.3216 -0.2771  3.0542  6.4785 -1.4086
#> 
#> (1,5,.,.) = 
#>   1.0796 -3.4794 -7.4299  6.1925  5.7134
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]