Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    3.5830   4.0371   6.2477  -0.5477   5.2215
#>    9.0030   4.8197   7.3039   3.8311   5.3702
#>    3.9370  12.0337   5.6565   7.8856  -3.5322
#>    9.2780  -2.3936   9.5940   7.0115   4.7347
#>    6.4558  -2.6480  -4.1760   5.5890   5.6349
#> 
#> (1,2,.,.) = 
#>    1.0521  -2.8220  -3.0031   0.4209   2.8595
#>    6.4411  -3.6989   2.7058  -3.4597   0.8128
#>    4.4952   4.4366  13.9229   6.9856  -0.5876
#>    2.8530   5.8223   4.0957   7.7056  -1.5304
#>   -4.3648   0.6605   5.5931   2.8435   0.9052
#> 
#> (1,3,.,.) = 
#>    3.8416   1.1644   4.0712   1.8013  -2.0351
#>    1.4089  11.4452   1.2058  -3.9895   2.9582
#>    3.9882  -7.0430   2.5893  -5.1422   3.9746
#>    0.1858   2.3610  -4.3157  -2.5372   4.3176
#>   -1.0645  -6.4163  -2.5555   8.1971   0.3041
#> 
#> (1,4,.,.) = 
#>    3.6776   0.4998  -1.6359   3.5891   0.4342
#>   13.0478   8.0104  -5.7118   3.9770   1.1634
#>    6.0140  -7.4293  -4.8095   2.2626   4.7907
#>    5.2028 -20.8457   5.0292   2.3311   0.9925
#>    4.5141 -13.6540   1.9678   8.2274  -5.7767
#> 
#> (1,5,.,.) = 
#>   0.5382  6.6247  2.0452 -0.0497  1.7229
#>   5.9533  1.0996  2.9471 -4.9553  2.5277
#>   7.9540 -4.6320  1.1074 -5.4064  1.2904
#>  -2.5327 -3.8502 -18.7137  7.8386  2.2678
#>  -2.9637 -2.7155 -2.5836  2.9719 -0.0230
#> 
#> (1,6,.,.) = 
#>    1.3182   0.1397  -4.1780   0.9632  -0.6746
#>    4.9640  -3.7389   4.8831 -13.2527  -1.6518
#>   13.1405   2.5706   2.9157  -2.7697   5.2985
#>  -11.7518  -9.4566   7.9261   8.7331  -0.3074
#>   -7.3916  -0.2266   2.1942  -3.1042  -4.7732
#> 
#> (1,7,.,.) = 
#>   0.0609  3.8808  1.7234  0.3206 -0.9506
#>  -1.5761  0.6021  1.9632 -2.1541  4.6090
#>  -5.2351  0.9825  7.4537  7.2473 -1.3711
#>   1.9238  8.5730 -1.1138  0.6922 -1.1933
#>  -4.3378 -3.9142  1.0620  2.0465  2.6894
#> 
#> (1,8,.,.) = 
#>   -0.0238  -3.1680   2.3766  -0.5191  10.5372
#>    0.0309 -18.1001  -4.6009  -2.8785   9.2018
#>   -3.9665   3.6091  -2.8742  -2.8685  -4.1238
#>   -5.0814   5.9534 -10.5626   4.0559  -2.0837
#>  -10.0936   0.5049   2.8487  -6.4424  -3.5448
#> [ CPUFloatType{1,8,5,5} ]