Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -2.4315   4.5352  -7.0916   8.4465  -5.8452
#>   -2.4718  -3.1608  -7.1100  14.2558  -0.1929
#>    1.5696   1.4572  -8.8068   5.5233   4.2215
#>    3.9130   1.0609   3.2017   3.5378  -0.9063
#>    3.0146  -1.3507  -6.2017   0.9899  -5.1071
#> 
#> (1,2,.,.) = 
#>    0.4461   0.4900 -11.5738   4.7658  -3.5116
#>    6.4207  10.2649  -9.1483   4.7267  -1.0044
#>    2.7232  -8.4452  -1.6626   1.1238  -0.5700
#>    4.2320  -2.4048   3.9287   1.6532   2.8501
#>   -4.3174   1.2986  -3.1263   1.8407  -1.7075
#> 
#> (1,3,.,.) = 
#>   -5.9704  -2.9080  -6.6026  11.2179  -3.6628
#>   15.3604  -0.1411   0.7604  -2.1912   0.8260
#>    1.7379   1.5197  -2.9540  -7.8091   4.3014
#>   -0.9034   6.1203   0.5667   4.0052  -3.6163
#>    0.0469  -0.7643  -4.8190   3.1820   0.6423
#> 
#> (1,4,.,.) = 
#>    7.2800  -5.0039   3.3524  -2.7327   0.3321
#>    0.6492 -10.8952  -4.1890   4.7032   0.2912
#>   -1.2456   7.1874 -18.9185   8.5848  -3.4377
#>   -6.8818  -4.7911  14.2964  -0.9805   5.8744
#>    1.3481  -0.8431   6.9982  -2.8262  -5.1288
#> 
#> (1,5,.,.) = 
#>  -0.9701 -5.2217 -1.1348 -1.1898  5.0484
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]