Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -3.3679  2.9666  3.1715  6.0771 -4.6065
#>  -1.2926 -5.3117  0.1093  2.0708 -1.6098
#>   0.7621 -0.2727 -2.2241 -7.3459  6.8174
#>   6.7619  6.4666 -0.3864 -4.5837  5.0686
#>   2.3908  8.2094 -2.2283 -1.9428 -0.4834
#> 
#> (1,2,.,.) = 
#>  -1.9980  2.4082 -7.1165 -1.0292 -1.8509
#>  -3.4819 -7.5966 -15.3588 -11.3631  4.0112
#>  -0.9367  9.3054 -2.6572 -3.9487 -4.4108
#>  -10.4166  9.4226  6.2544 -6.7679 -6.9250
#>  -0.4155  1.0991  0.4273 -4.0856 -1.6966
#> 
#> (1,3,.,.) = 
#>   -8.0206  -3.8038   4.1851   1.5000   2.9721
#>   -2.6254   6.4887  -2.8812   8.2359   5.5960
#>   -3.9324   8.2924  -6.2227  -1.3236  -1.6352
#>   -8.3438   5.6812   7.8713  -7.0699  -4.2400
#>    1.2384  11.6769   6.4583  -4.9650  -0.2653
#> 
#> (1,4,.,.) = 
#>   -3.4403  -0.8321  -4.9125  -0.1510   2.7791
#>   -0.6961   3.9940  -7.7378   0.8765  -8.9887
#>   -2.2235   4.6197  12.4998  -5.5607  -6.2103
#>   -1.9883  -0.8827  13.0800   1.2648  -4.4376
#>   -5.2054  -0.7117  -1.1006   2.5114  -2.9889
#> 
#> (1,5,.,.) = 
#>   -2.2401  -1.7484   0.3901  -1.6106  -2.1032
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]