Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    3.6313   0.9712   2.3766  -2.5321  -0.5062
#>    6.7055   0.7744 -10.6312  -0.4361  -6.0090
#>   12.4095  -0.7520  -1.6073  -4.8157  -1.4992
#>    0.2855  -0.1729  -1.4494  -1.6743   5.6827
#>   -0.9603   5.5485   1.1078  -1.2579  -4.1413
#> 
#> (1,2,.,.) = 
#>    2.4407   1.7764  -4.8770   2.8870   0.8398
#>   -0.2300 -10.6125   6.5995  -4.4652   2.3182
#>   -1.1398   4.5009  -5.6917   0.3123   3.5520
#>   -5.6349  -6.0029  11.4020   2.2793   3.6665
#>    5.1905  -4.0735  -5.7869  -1.3690   2.3032
#> 
#> (1,3,.,.) = 
#>   2.0463 -3.6153  0.2938  0.6750 -6.5398
#>   0.9406 -2.7397 -0.7686 -1.1503  1.8989
#>   8.5751  2.9153  2.7099 -1.0235  4.2506
#>   1.1342  3.0443 -16.0698 -1.0888 -0.2703
#>   7.8649  2.3529 -0.1057 -5.5180 -3.1425
#> 
#> (1,4,.,.) = 
#>   -1.2861   6.3798   1.1788   4.1164   2.7782
#>   -1.7474  -3.0357   2.5070  10.0196   3.6855
#>    3.6764   8.6730  -5.4091  -5.9236  -3.2855
#>    5.1156   1.0548  -1.1537   8.2320  -0.3128
#>    3.2641  -0.3819   2.1593  -4.5201   2.4471
#> 
#> (1,5,.,.) = 
#>  -1.9692 -1.2235 -0.7575  4.2505  2.6145
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]