Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   2.5742  5.2684  5.1131  0.1072  4.4727
#>   1.4220  2.2071 -0.1175  1.9779  2.7837
#>   2.6315  1.5461  4.0698 -5.5203  5.7865
#>  -2.7563  5.9852 -3.8826  5.4072 -3.6466
#>  -5.4900  1.3381 -1.0520 -1.9547 -1.3984
#> 
#> (1,2,.,.) = 
#>   -6.1065  -6.7691  -2.9044  -2.6756  -1.1731
#>   -2.7568  10.2578  -3.4495   0.2451  -6.9826
#>   -6.5293   6.2606   0.4536  -2.5525  -6.8009
#>    3.9933  -3.3236   4.7668  -4.4729   4.9740
#>    3.8518   1.0731   7.9505  -6.3419   8.5843
#> 
#> (1,3,.,.) = 
#>  -4.4538  5.2011  3.7496 -4.4704 -0.5740
#>   5.1474  5.4948  3.8369  1.9545  5.0972
#>   1.7735 -2.3037 -5.2420  0.0954 -3.2529
#>   0.8978  2.7566  1.0107  2.3851  3.3026
#>   5.9173 -1.7923 -0.5261  3.0707 -4.0524
#> 
#> (1,4,.,.) = 
#>    5.5878   2.2401   2.7051  -4.3786  -2.9752
#>   12.2967   8.4044   0.3852  -0.8902  -0.0830
#>   -0.4530  -0.3076   4.5759  -3.8780   1.5491
#>   -8.1302  10.8183  -2.0347   9.3794   3.5162
#>   -7.3510   2.0813   4.8741   5.6130   1.4835
#> 
#> (1,5,.,.) = 
#>   -4.1782   4.1131   2.1735  -5.1595   6.3547
#>   -0.8910  -1.7505   1.9447  -4.4840   6.5722
#>   -7.1756   1.4562   3.2304   0.0697  -0.3015
#>   -4.1186  -1.5269  -2.3578  11.4718  -3.3637
#>   -2.7980  -8.3835   2.2381   4.6926  -2.5572
#> 
#> (1,6,.,.) = 
#>  -1.1925  0.7342  1.6905 -4.6844 -1.1512
#>   0.2387 -3.2609  2.9723  1.4925 -4.5190
#>   0.6702  1.1957  3.0847 -0.6850  7.8634
#>  -4.8338 -5.1935  2.2406  3.7941 -2.5951
#>  -2.6505  9.9269 -3.3626 -1.0039 -4.3999
#> 
#> (1,7,.,.) = 
#>    0.3714  -2.4722   2.7040  -3.3349  -4.2082
#>   -0.4702  -0.1059   6.1954   3.8960  -6.5797
#>   -0.3978  -1.0095   8.3945   0.7777  -4.3459
#>   -1.0909  -0.6383  10.6762   5.6717   4.3207
#>   -1.0805  -3.0707  -5.9113  -0.1518   4.8932
#> 
#> (1,8,.,.) = 
#>   -2.0201  -0.5472   0.8377   0.7629   1.2206
#>   -3.8939   4.2666   1.7900  -3.9845  12.5486
#>    1.2086 -11.2016  -3.0396   2.3658  -5.5870
#>    2.3463  -3.5790   4.3266  -4.0349   4.4791
#>   -4.5389   0.1278  -0.1564  10.1393   1.7231
#> [ CPUFloatType{1,8,5,5} ]