Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -3.2791   2.9723   2.7802  -8.2573  -6.1997
#>  -11.9581 -17.6202   2.0937   1.9514   6.8119
#>   -2.6790   7.8249  -6.9486  13.3818   3.9966
#>   -3.9037   0.9637 -17.7070  -1.0499   4.3087
#>    1.8265   3.4247  -0.2954  -2.1662  -1.0478
#> 
#> (1,2,.,.) = 
#>    2.2387   4.4018   0.7422  -0.2444  -1.2756
#>   -1.2736  -3.1208 -12.8632  10.2156   0.1656
#>   10.7444  -5.8585  -0.8256 -16.4424  -1.6147
#>   -1.0910   3.4174  -3.8140 -10.9209  -6.5373
#>   -5.0537  -3.6082  -3.9836  -1.0814  -3.8564
#> 
#> (1,3,.,.) = 
#>  -1.8754 -4.1965 -2.8439  2.3426 -0.4208
#>  -0.0265 -5.9815 -6.8247  1.7228 -1.2321
#>   4.9437 -3.4458 -0.9158  1.7125  7.8090
#>   1.3930  6.8975  5.2232 -8.5306  0.2179
#>  -3.3708  1.9777 -0.4319  1.8550  1.1297
#> 
#> (1,4,.,.) = 
#>   1.0775  8.1038 -9.1078 -3.1146 -5.2902
#>  -3.7460 -0.1704 -1.2519 -0.0819  0.4985
#>   7.9857 -10.4972  7.4374  2.8350 -5.1023
#>   1.2928  2.7600 -8.4077 -4.8944 -1.7265
#>  -7.7582  5.0916  4.7973 -7.9931  1.5644
#> 
#> (1,5,.,.) = 
#>  -5.9783 -9.5910 -4.0822 -0.8106  0.1133
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]