Conv_transpose2d

Conv_transpose2d

Arguments

input
NA input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
NA filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
NA optional bias of shape \((\mbox{out\_channels})\). Default: None

stride
NA the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
NA dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
NA additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
NA split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
NA the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=None, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See ~torch.nn.ConvTranspose2d for details and output shape.

.. include:: cudnn_deterministic.rst

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   -1.0514   4.0431   5.3680   0.4996   0.6807
#>   -2.2791   3.4717  -9.2100   2.1827  -0.9895
#>   -8.7717  -1.1484   0.2236  10.5655  -7.9790
#>    3.4481   4.1267   1.9447  -3.6601  -2.1014
#>    3.1070  -0.9492  -1.0022   4.1038   3.1708
#> 
#> (1,2,.,.) = 
#>   3.8769  1.0065 -1.7453  7.8551  1.1373
#>   1.5910 -0.6408  0.0578  0.3598  2.9826
#>  -2.4286 -6.3642  8.1915  3.7459 -6.1005
#>  -0.9822  3.3808 -9.7499 -1.4318 -1.9887
#>  -1.2174 -7.7774 -4.6879  4.6992 -0.3623
#> 
#> (1,3,.,.) = 
#>   -0.9979  -3.4498  -3.6999   2.2445   2.7840
#>    2.1584  -2.2082  -1.5000   1.7348  -4.5991
#>   14.2319  -0.5299   8.7039  -6.1999   3.8417
#>   -4.0469   1.8079  -6.1525   5.3468   3.7647
#>   -2.3751  -4.6096   6.3424  -1.9923  -5.4673
#> 
#> (1,4,.,.) = 
#>    1.6146   2.2870  -7.7727  -0.5212   1.2332
#>    3.8877 -11.1170  -2.1337  18.7782  -2.8248
#>    7.9747  -6.2239   6.6249  -8.7275  12.3155
#>    2.1709   2.9160  -1.5355   1.4338   1.7243
#>   -1.2307   2.9260   6.1498  -3.0348  -5.3836
#> 
#> (1,5,.,.) = 
#>   2.8414 -5.4122 -1.7579  3.6993  3.3308
#>  -3.7484  5.5068 -3.1213 -5.1444 -4.9850
#>   2.1524 -7.8374 -4.9869  9.7475  4.5635
#>  -2.7844 -1.6397 -5.1861 -3.9104  6.2100
#>  -7.9028  2.6999 -4.1059  8.2526 -2.6561
#> 
#> (1,6,.,.) = 
#>   -6.3073  13.8461  -1.2210  -4.8289   1.8385
#>   -6.8310  13.7232   9.1193  -3.5035  -3.3996
#>    7.6498   2.7539   0.1421  -8.5598   4.6140
#>   -7.7794  -3.5150  -3.2721   6.2342  -5.6870
#>    0.7614   6.2618  -8.7980  -8.8184  -0.8238
#> 
#> (1,7,.,.) = 
#>    6.6540  -5.3689  -2.0845  -3.5825  -0.2029
#>   -0.8108  -9.5855   3.2040   9.2180  -0.0728
#>   -5.6512   0.8883   8.3345   2.0343  -0.5770
#>    7.8277  -0.2300  -8.6879  -8.4446  10.1034
#>    2.0860  -6.7967  -8.8847   9.0482  -3.5302
#> 
#> (1,8,.,.) = 
#>  -2.8995  7.8823  5.0617  0.5166 -2.8131
#>   3.7143  1.5723 -0.0380  2.8057  1.6559
#>   1.1504 -5.3717  7.8838  7.3967 -8.2488
#>   7.2704  7.1180 -2.2714 -6.7369 -1.3559
#>   8.4336 -1.3131  6.6604 -13.5235 -0.8044
#> [ CPUFloatType{1,8,5,5} ]