Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   0.2891  0.5457 -5.9133  3.7804 -0.9630
#>   2.3734 -6.9930  2.8144 -2.1310 -0.6533
#>  -6.5731 -1.3384 -2.7784  4.3247  3.8550
#>   0.3010  1.5800  8.9416 -8.1878  6.5816
#>   5.1898 -0.1257  3.8383 -2.9503  2.0873
#> 
#> (1,2,.,.) = 
#>  -2.6883 -9.6173  3.4463 -1.3967  3.3831
#>  -1.6244 -5.9892  2.9017  0.8745  1.6956
#>  -3.4629 -8.6816  7.2009 -6.4997  1.1174
#>  -1.3975  2.2294  4.5502  2.2228  9.3941
#>  -2.9310  1.8624 -0.2969 -10.9033  8.9633
#> 
#> (1,3,.,.) = 
#>    1.1703   4.4868  -1.7190  -6.4183  -2.5418
#>   -0.9867 -15.5343   2.6530  -7.5318   2.9611
#>    0.5977  -4.5430  -2.7538  14.5138   6.0484
#>   -2.9867  -0.5224  14.1638 -13.8417  -4.1324
#>   -6.0083  -6.5746   4.5078   8.2631  -7.0014
#> 
#> (1,4,.,.) = 
#>  -1.2017  0.2527  6.6079 -2.5913 -3.7673
#>  -1.8525 -0.9096 -1.5930  0.3855 -3.3955
#>   6.8799 -12.1058  0.6124  0.1078  1.0066
#>  -1.9277 -0.6030  6.8010 -5.6500  6.7608
#>  -3.0431 -2.9924  6.0519 -1.8755  2.6065
#> 
#> (1,5,.,.) = 
#>    0.5528   0.3324  -4.2958   2.0203  -0.4328
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]