Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   3.6642  2.9325 -2.9688  1.4417  2.2564
#>  -0.8582 -7.4183 -2.2456  1.6820 -0.1615
#>   0.3423  3.2211 -6.4123  0.8171 -15.6567
#>   6.8774  2.0773 -12.0303  7.7584 -1.8290
#>   0.0727  2.3813 -5.2719  1.1129  1.7507
#> 
#> (1,2,.,.) = 
#>   -0.4307   7.7742   4.3687  -5.9887  -4.6133
#>    1.4449 -11.6321  18.6994   8.1957   1.4896
#>    0.8718  -3.4946  -1.9533   1.4125   4.3297
#>    3.7707   2.6872   9.0623   8.2528   0.7782
#>   -1.7172  -6.2256   5.3806   5.4255   0.7388
#> 
#> (1,3,.,.) = 
#>   1.0820  2.2889  7.7265 -1.2999  3.5496
#>  -7.5134  2.9513  4.4938  2.7323 -2.8488
#>  -2.1710  0.5782  4.6042  6.6828  5.8057
#>  -8.5187 -3.6870  1.7425 -9.2553 -1.3452
#>  -1.3645  4.3928 -0.6636  1.9619 -1.4102
#> 
#> (1,4,.,.) = 
#>    4.2463 -14.0995   5.7270   2.3877  -4.2968
#>  -10.5574   8.2407  -3.3519  -5.9392  -2.3548
#>   -6.5338   3.1906  -1.3755  -5.3092   4.4688
#>   -2.6144   5.8329   1.6283  17.9841  -0.6422
#>   -4.4603  12.1120  -4.4431   2.7306   1.3190
#> 
#> (1,5,.,.) = 
#>   -2.3900  -7.0375   7.3465   0.6894  -1.9509
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]