Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    8.6880   6.4520   0.4849  -0.9313   5.0898
#>   -3.1885 -15.1802   1.1711 -10.8686   0.3839
#>   -2.3026   2.1260  -4.5588  -2.0009  -0.9407
#>  -14.9228   0.9529 -17.1317  -2.4316  13.8492
#>    0.0335  -5.5128   3.6683  -9.8895   9.4107
#> 
#> (1,2,.,.) = 
#>   -1.5622   0.0145   3.3720   0.9241   0.3013
#>   -4.1360  -6.8228  -1.8490   1.7247   6.4141
#>   -9.9914  -8.6259  13.6879   5.3614   3.5602
#>  -10.3281  -2.3035  -1.3810  11.5148  -5.3686
#>    2.1723   4.7213   1.9077   1.3933  -1.8024
#> 
#> (1,3,.,.) = 
#>   -4.3449   0.1392   3.6072   7.4128   0.1140
#>    2.5960   6.7845 -12.5278   3.8490   1.0666
#>    5.0654  -0.8583  -3.0937   2.0667  -2.5347
#>   11.1458  -0.2931  -1.6121  -5.1487  -0.7654
#>    2.2455   8.0431  -3.2398   3.9133  -1.3361
#> 
#> (1,4,.,.) = 
#>   -7.1937  -5.2184   3.7235  -1.5811  -3.6964
#>   -1.1349   1.1750  -3.2747   4.6600  -3.3421
#>    4.4810   1.3147   6.6222  -3.0450   4.1235
#>   -2.0939   7.0837   2.7381  11.5841   0.7191
#>   -5.5508   8.5405  -3.5609   2.9087   3.7153
#> 
#> (1,5,.,.) = 
#>  -4.0844 -1.0616  6.1246 -4.6169  0.3025
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]