Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -2.9663 -1.2733 -3.9435 -0.5276  2.6750
#>  -1.2722  1.0069 -5.7263  2.7787 -0.1404
#>  -1.7860 -3.8682  4.8713  1.1316  3.8511
#>  -0.0010  5.3181 -0.6754  2.5730 -2.4711
#>   0.5885  4.9659 -3.5166  4.5335 -3.7058
#> 
#> (1,2,.,.) = 
#>  -5.2250  0.2726  8.2119 -6.4099  5.9053
#>   4.6700  5.1411 -1.5878  3.8120  0.1009
#>  -3.5450 -5.0390  9.2086 -2.9578 -0.5509
#>   8.0991 -14.7839 -3.2309 -12.1295  2.1448
#>   0.1715  2.9083 -1.4596 -1.0306 -2.7095
#> 
#> (1,3,.,.) = 
#>   3.1749  2.3139 -8.2462 -6.3721  0.0100
#>   0.8951 -3.3390  0.3912 -8.3079  1.3706
#>   3.5525  3.0387  6.1074  6.3115 -2.0431
#>  -3.2812  2.3111  2.1700  3.3244  4.8020
#>   5.2995 -1.5489 -1.4167 -6.5637 -0.9624
#> 
#> (1,4,.,.) = 
#>  -10.0481   7.0235  10.5055  -2.0748  -3.7513
#>    8.1387  -0.4205  -3.0138 -10.8628   5.5135
#>    2.9024  -5.1426  -1.6595   4.3741  -5.7961
#>    4.4890 -14.6718   5.5729   4.6866  -0.7137
#>    1.1518   3.0874   0.7841   0.3746   0.9620
#> 
#> (1,5,.,.) = 
#>    0.9123  -8.0814   1.0808   7.4471   0.6151
#>    2.1226   2.7764  11.3909   4.8650  -4.3402
#>  -11.2861   4.5080  -7.8865   5.5343  -9.9356
#>   -1.9028  -1.0153  -3.6493  -7.0338  -8.3456
#>   -3.5633  -8.6804  -7.8108   6.3765   2.8930
#> 
#> (1,6,.,.) = 
#>   -0.5542   1.5379   3.6290  -4.7070  -6.1971
#>   15.2333   1.0645   0.0866  -3.6305   0.7299
#>   -0.0833 -10.8658  -6.3612   0.5463   2.2777
#>   -4.8246   2.8323  -2.6610  -1.3371  -0.6676
#>   -5.5904   6.6461   2.8359   5.6594  -0.5057
#> 
#> (1,7,.,.) = 
#>   2.0289 -0.2122  4.9581  0.2454  1.7928
#>  -4.9431 -0.3061  5.8549  0.7861 -2.7473
#>  -3.8568 -3.2374 -1.5201 -1.2617 -7.8991
#>   1.0549  1.0301  3.8253  1.0461 -6.5575
#>  -3.1756  2.8057 -2.0838  7.5251  3.8319
#> 
#> (1,8,.,.) = 
#>   7.6274  4.6864 -5.2624  0.0808 -3.7774
#>  -0.0753 -18.2773 -7.6814 -6.3614  1.1808
#>  -0.9235  9.3308 -1.1801 -6.3923  2.6811
#>  -0.3068  9.7232  2.0825  5.2834  8.0793
#>   5.7741  8.8292  2.7251 -3.5073  0.5215
#> [ CPUFloatType{1,8,5,5} ]