Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    4.7384   0.7926  -2.1987  -2.4575   7.9759
#>    2.2843   4.4220   3.2338   8.1882   3.4721
#>   -1.3857  10.9831   5.5591  -0.3265   4.8525
#>   -0.2685  -3.1097  -6.7087  -1.7067  -3.9130
#>   -3.9001  -3.8791   4.5543   8.1960   3.8248
#> 
#> (1,2,.,.) = 
#>  -4.2644 -4.3927 -4.1769 -0.4581  8.7706
#>  -1.9405 -1.1975  7.2938  4.1093  4.3028
#>  -5.0472 -4.2787 -1.2145  5.7701 -1.3533
#>   1.0268  1.9248  8.4090  9.7852 -4.3485
#>  -5.2074 -9.2789  3.2759  4.4067 -0.2722
#> 
#> (1,3,.,.) = 
#>   -1.5871  -3.5784  -5.8415   3.3153  -9.5823
#>    0.9377   2.2042   3.7778   3.9576  10.6895
#>    2.5528   2.2654  12.3054   3.2988  -0.6133
#>    7.5674   1.6717  -4.6378   3.7701  -5.9743
#>   10.9825  -5.3059  -5.6484 -10.8510  -7.4806
#> 
#> (1,4,.,.) = 
#>    1.4560   0.3953   2.9381  -7.3878  -3.4939
#>    3.1997   0.1728   2.1604  -1.6077  -6.8471
#>    6.2945  -0.4677  -4.7303  -2.6133  -3.4355
#>    4.2561  -4.0163   8.1361  11.3576   6.8931
#>   -7.9824  -1.0247  -3.0928  11.1896   7.1729
#> 
#> (1,5,.,.) = 
#>   -3.9531   4.7856   7.5941   4.1006   4.8316
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]