Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   3.5851 -2.6604 -1.6711 -2.2592  1.8697
#>   7.1247 -0.0730 -12.6205  7.9452 -0.2441
#>   0.2860  3.5489 -1.6076 -6.4871 -4.1346
#>  -2.2569 -2.1704 -1.3510 -7.1092  8.3522
#>  -5.3819  0.6729 -2.4536  1.1872  1.3975
#> 
#> (1,2,.,.) = 
#>   -4.6164  -1.8800  10.2739   0.7210  -2.0015
#>   -4.6396  -7.5246   2.8153  -0.7363   1.3365
#>   -3.0254  -2.7832  -7.4010   5.6266   1.3492
#>    2.2085   8.2924  -3.1305   6.7899  -3.3799
#>   -1.2124   4.4257  -4.1029   7.0700   1.5269
#> 
#> (1,3,.,.) = 
#>    8.3913  -8.8739   0.5830  -7.3261   4.9202
#>   -3.2886  15.8683  -1.0020   1.3420   3.3884
#>    4.4809  -0.0565  13.4985  -1.3872   6.9427
#>   -7.2364  -2.4401   7.6876  -5.5114   3.6920
#>   -2.5270  -5.1960   0.3307   0.9221   2.2397
#> 
#> (1,4,.,.) = 
#>    4.7426  -0.7610  -1.8560  -1.8470  -2.8272
#>  -10.5372   9.4593   0.2497  -5.1779  -5.1656
#>   -0.4578  -2.7113   8.0699   0.3299   7.1559
#>   -1.0355  12.5291   3.3884  12.2725  -3.3882
#>   -2.4360   0.9004   2.8193  -2.4730  -0.4705
#> 
#> (1,5,.,.) = 
#>  -6.8252 -5.5225 -4.9462  6.3280 -1.4432
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]