Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   2.5042  2.6701 -8.4941  9.1701 -1.5770
#>  -0.3569  4.7657  7.0070  5.4644  1.0749
#>  -6.5254  4.6645 -5.9239 -6.7211 -1.7606
#>   5.4716 -4.1012  3.0077  5.8727  1.0506
#>   4.9693 -2.4692 -5.7955  0.6366 -1.3381
#> 
#> (1,2,.,.) = 
#>  -0.6727  1.1892 -7.7068  7.2204  1.2124
#>   1.3811  5.1772 -5.3260  0.7055 -1.8171
#>  -2.4070 -1.0409  7.2663 -1.8844 -3.2916
#>   5.0613 -0.7986  6.1931 -10.0075 -1.3620
#>   7.3021 -12.1972 -5.2211  4.5314  2.7377
#> 
#> (1,3,.,.) = 
#>   2.2527 -2.9857 -2.7110 -1.2545 -4.4697
#>  -1.2349  3.0192 -4.4844  2.1069 -3.2276
#>   0.3154 -0.8410  0.6534 -0.9080 -5.1307
#>   5.9283  6.3267 -5.5406 -5.0623 -1.5624
#>   0.3060 -6.4998  3.0873  4.3306  0.8258
#> 
#> (1,4,.,.) = 
#>    2.2096   9.6327  10.0995  11.6111   7.6784
#>  -11.7693   6.2498   3.4589  -8.9430  -6.7583
#>    2.5078   4.0445  15.9263   6.3610  -1.5057
#>    7.5904   0.8780  -4.7079  -4.3029   9.0509
#>   -1.8794   0.1489  -0.8613  -1.9881   0.6362
#> 
#> (1,5,.,.) = 
#>   -3.9413   6.3101   2.1634  -0.1234  -5.1366
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]