Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -6.9356   2.4915   6.2579   3.4390   2.9287
#>   -5.9464   3.7967  -6.6683   2.6020  -5.9623
#>    4.8178   7.7267  -6.1737  -6.6446  -0.9066
#>    3.7255   1.4272   1.9795  11.1045  -1.5331
#>   -3.6124  -5.6232  -8.2647   0.8605  -3.4535
#> 
#> (1,2,.,.) = 
#>  -3.0478  5.5642  5.0106  4.2346 -1.1507
#>  -1.7135 -6.8849  3.7111 -5.1792  4.0476
#>   3.5987 -3.0123 -3.3057  0.3333 -3.4123
#>  -4.9122 -3.5271 -1.7456 -4.7096  2.4169
#>  -0.3884 -5.0619  9.5937 -5.8885 -2.1580
#> 
#> (1,3,.,.) = 
#>  -1.6406 -6.3060  2.5195 -2.2325  1.7662
#>   4.1696  0.5463  3.3858  3.9947 -1.8705
#>   6.6884  2.9788 -16.2672  1.6108  1.2450
#>   3.2374  2.5308  2.1981  1.9955 -4.6579
#>  -3.0655  3.8688  2.2215  8.3576  3.7701
#> 
#> (1,4,.,.) = 
#>   11.6416  -6.8388  -1.9831  -5.7536  -7.6201
#>   -2.9189   2.5916 -12.7909   0.5102  -0.7611
#>   -5.2958   6.8256  17.4315   4.1299  -5.2592
#>   -1.9002   6.4395  -3.4345  -2.7750  -0.6304
#>   -4.8324  -1.0977  -2.7966  -1.1973   0.5909
#> 
#> (1,5,.,.) = 
#>    2.8380  -4.0079   4.0055  -4.9668  -0.9203
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]