Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   1.5068  1.7804  1.4758 -6.4905  7.9177
#>  -3.1193  3.9779  8.1286 -15.8510  1.4718
#>  -0.2243  3.7642 -10.8167 -8.5549 -4.1442
#>  -2.6042 -3.8256 -7.4095 -4.9577 -1.4879
#>   4.3131 -7.6089 -6.2900  5.2631 -4.7786
#> 
#> (1,2,.,.) = 
#>    0.5632   8.1676  -5.5867   5.9871   1.0162
#>    1.2019  -4.8391  -1.4160  12.4501  -5.0216
#>   -3.2833 -10.5750   4.9966   6.0268  -1.8358
#>   -3.8690  -3.9683  -5.0940 -15.7750   6.5646
#>    1.5451  -1.6895   6.5307  -2.8567   3.6501
#> 
#> (1,3,.,.) = 
#>    9.3918  -5.2506  -1.3503  18.2278 -10.7920
#>    4.3736  -0.4130   8.2515  -0.0637  -1.9479
#>   -4.6206  -2.7898   0.6722  -9.4183  -1.4064
#>  -10.7983  -1.2334   0.6931   1.7384   5.2484
#>   -2.8866   4.6340   2.7399  -0.0962   4.5021
#> 
#> (1,4,.,.) = 
#>   3.5758 -5.1499 -0.8380  5.2772 -1.0343
#>  -3.3931 -0.2816  4.9985  1.5631  0.3749
#>   0.9404  6.1371  6.0317  4.3386  1.9513
#>  -3.8023  2.3387  0.5624 -1.5748 -1.7538
#>   1.7699  2.7989 -6.8735  1.2757 -1.4095
#> 
#> (1,5,.,.) = 
#>    4.2291  -6.3548   5.8123  -2.3062   2.0639
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]