Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -3.4161  -9.2479   6.3997  -2.5352   5.7947
#>    6.8936   2.9031  -3.3045  -3.1655   0.3911
#>   -3.0666  11.4662   7.7931  -6.6278  -1.8758
#>   10.2267   2.9669  -7.2771  -3.3508   0.3131
#>    0.1387   5.0212   1.2539  -2.7500  -1.7509
#> 
#> (1,2,.,.) = 
#>  -4.8490  2.1785  2.5679  4.8057  0.0533
#>   2.3900 -15.4260  3.1641  3.3056  0.9366
#>  -6.9489  5.1260 -3.3918 -0.7289 -0.0564
#>  -3.6279  2.2872  1.2408 -1.8778 -2.1488
#>   8.6355  1.7410 -4.7252 -8.3949  4.5539
#> 
#> (1,3,.,.) = 
#>   -2.8982  -0.5689   4.4019   9.6861   4.1031
#>    8.5916   4.0298  -3.5535   1.3801   4.7029
#>  -10.2674  -4.9697  -3.9133   1.9553   1.5376
#>    3.5035  -6.6098  -3.9261  -6.5785   0.6623
#>   10.3982   1.2400  -2.4087  -3.0606  -3.3159
#> 
#> (1,4,.,.) = 
#>   1.9034  3.2719  4.5490  2.7685  2.5976
#>  -8.9282 -1.0126  0.9092  6.1112  7.0410
#>  -0.8881  0.2458 -0.9249 -0.9699  7.7548
#>   4.8522  2.7041  3.3026  3.4468 -0.0787
#>   4.6739 -5.6216 -7.2876 -0.9867  2.9681
#> 
#> (1,5,.,.) = 
#>   10.3447   6.0784 -10.2153   3.9688  -3.1252
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]