Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -4.6746 -3.5147 -1.3725 -1.1722  2.3773
#>   3.2738  5.9347  0.4147 -0.1371  1.0476
#>   4.8528  0.8768 -1.7607  4.6776 -0.0095
#>   5.9633  0.3581 -0.0274  0.5424  1.1303
#>  -1.2942 -2.0834 -0.3010 -3.5026 -1.5328
#> 
#> (1,2,.,.) = 
#>  -1.3862 -6.1558  2.6933  3.9860  6.2549
#>  -0.0321  5.5121  9.9993  1.4770 -4.6824
#>   5.3351  8.5785 -5.6292 -0.4272 -6.6760
#>  -0.8114 -0.1485 -8.6209  3.5435  6.7186
#>  -1.6726 -2.6255 -0.8642  4.7024 -2.0861
#> 
#> (1,3,.,.) = 
#>  -0.7070 -3.7097 -6.2265 -0.1817  1.5346
#>   4.2831 -7.8837  1.3971 -6.0346 -1.3844
#>   4.8929  8.4912  3.5570 -8.7070 -0.0520
#>   3.2301  5.6890 -6.0916 -6.4779 -3.4228
#>   2.9522 -5.4797 -3.4238 -1.4997 -0.6513
#> 
#> (1,4,.,.) = 
#>   1.2210  0.1244  8.2731  1.4810 -6.1419
#>  -3.4222 -8.5198  0.8784 -4.3376  0.4562
#>  -2.4724 -3.4927  1.0399 -0.3115  5.0619
#>   0.6551 -7.9521 -5.9080  7.0757  1.3159
#>  -1.7808 -6.9507 -0.5140  9.8038  5.1851
#> 
#> (1,5,.,.) = 
#>  -1.0840 -2.0746  2.9626  7.0440 -0.2763
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]