Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.0580  -8.9025  -1.0059   5.1985   4.8346
#>    0.9958   4.8331  -4.8175   5.4156   0.5081
#>    5.2330   5.8887  -6.6339  10.2978  -2.0729
#>    2.8933   0.2477  -2.7429  -2.7576   0.7356
#>   -1.0691  -0.0328  -4.0752   9.4789  -1.7984
#> 
#> (1,2,.,.) = 
#>    8.9441  -1.2216   1.7627   0.5854  -0.9526
#>   -3.1211   0.7416   5.3235  -2.4543  -4.5014
#>   -0.1228  -3.0456   7.3094  -0.1402  -7.5177
#>   -1.1391  -2.8802  -1.6372  10.3168  -6.4568
#>   -1.4541  -0.8985  -4.7890   0.4094  -2.7513
#> 
#> (1,3,.,.) = 
#>   -1.5761  -4.4319  -7.3741  -7.2035  -0.1505
#>    4.9020 -10.3782  -3.6412  -3.2491   3.2956
#>   -0.0312   1.6841  -3.0432  -6.2871  -1.1421
#>   -0.2106  11.7813 -13.1954   5.9380   2.0310
#>    2.2282   0.6086   2.3003  -4.1767   1.8610
#> 
#> (1,4,.,.) = 
#>    5.8419   0.1362  -2.2699   4.2198   7.7927
#>   -5.8556   3.4478  -5.0216   0.3907  -6.0524
#>    1.1757   0.9120  -7.2956   4.0744   1.1835
#>    1.4630  -4.0743  10.0253   4.3319   0.9464
#>   -0.8557  -0.3300  -0.8702   0.6122  -0.2548
#> 
#> (1,5,.,.) = 
#>    5.0773   4.0107  18.3574  -2.5151   0.8742
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]