Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -0.3587 -9.4636  4.3604  5.9074 -8.5312
#>  -2.1955 -4.4372  9.1189 -14.2233  7.5654
#>   1.6708  1.1212  1.6949  6.5445  1.3817
#>   7.9244  2.7399  1.5428 -11.4945  6.4228
#>  -0.0858 -4.2245 -1.7302  2.3111  1.4942
#> 
#> (1,2,.,.) = 
#>  -5.3083 -0.6444  0.2443  2.3384 -2.6127
#>  -1.7126 -3.7133 -3.2295 -6.0059 -3.8783
#>   5.3273 -7.5110  3.7469 -1.7196  3.6598
#>   4.2009  2.4839 -3.2982 -3.9202 -0.7316
#>   2.1259 -0.5211 -6.2176 -3.6114  3.8251
#> 
#> (1,3,.,.) = 
#>   3.0479  6.9722 -4.5222  4.6532 -6.6939
#>   1.2797 -0.8260  6.7681  1.0620  5.9633
#>  -5.2930 -1.7335  5.8547 -5.6485  0.3046
#>  -0.8069 -7.0388  4.5583 -11.0089  1.3193
#>  -5.3375 -4.3269 -1.4493  6.5234 -1.0325
#> 
#> (1,4,.,.) = 
#>   2.0453 -7.3899  7.6749 -11.7285  5.9552
#>  -3.1712 -0.8020  4.7253  2.8072  6.9458
#>   1.1342  2.2061  2.3460 -3.0194  4.4686
#>  -1.5326 -4.7021 -2.6114  3.3933  4.2548
#>  -3.5372  4.2040  5.0009 -4.3838 -2.0707
#> 
#> (1,5,.,.) = 
#>    4.1332   0.7551   5.9405   2.4641   9.2782
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]