Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -6.5761 -2.7675  0.8208  3.2312 -3.3132
#>  -5.5108 -4.1139  4.8988  7.9369  5.7381
#>   1.0398  0.1571  5.7261  8.9554  1.1924
#>  -8.1454  6.5308 -0.5635 -0.3777 -2.7926
#>  -0.3573  3.8956 -0.7745  3.6091  3.2280
#> 
#> (1,2,.,.) = 
#>  -1.9668 -6.6203  0.2308 -1.9647  3.0513
#>   3.5054  3.5605 -1.1012  3.1526  0.6225
#>  -15.5794  1.2814 -1.3505  1.1295  0.5342
#>  -2.5698 -5.5933  5.4319  7.9492 -0.2879
#>   2.2440 -4.3511 -8.9416  3.4743  0.6498
#> 
#> (1,3,.,.) = 
#>    5.7383  -2.2997  -4.5390   4.1226  -7.2794
#>    1.0353  18.0685  -0.3151  -0.3236   0.7541
#>   -2.1688  12.8502   6.8819   5.0667  -7.6135
#>   14.3357 -19.6079  -8.0643  11.1188  -4.4997
#>    1.4453   1.8622 -10.4914  -0.3013  14.6997
#> 
#> (1,4,.,.) = 
#>  -4.9426 -1.7498  1.6112  0.9164  0.6185
#>  -2.8617 -1.5796  3.1796 -12.9128 -2.5503
#>   3.2542  8.7907  0.5970 -3.7112  2.6147
#>   2.0900  6.9511  3.9473 -1.8068 -4.5877
#>  -2.8262 -3.1570  0.1749  1.9962  3.1459
#> 
#> (1,5,.,.) = 
#>  -3.6372  1.2872  6.9679 -0.7649 -1.1756
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]