Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   1.4336 -5.9808 -2.5927 -2.9712 -2.1510
#>  -2.0845 -2.1265  9.5589 -7.3413 -6.2091
#>   3.5209 -6.7076  0.0246 -11.2720 -0.7648
#>  -3.8194 -7.1525 -1.9168  0.2623 -1.8684
#>  -7.7454  5.5765 -5.4140  4.8306 -2.9292
#> 
#> (1,2,.,.) = 
#>   -4.6901  -2.5330  -0.6983   3.3891   9.6403
#>    4.8140   8.6600  -3.4251  -2.0217 -10.0911
#>    0.3360   1.7908 -10.4682  12.0366  11.1754
#>    0.8878   9.7293  -1.2978  -5.8024  -0.1395
#>    1.5745   6.9630   4.7072  -0.7681  -3.3525
#> 
#> (1,3,.,.) = 
#>   5.5752  3.8109  3.6409  3.6280  3.6291
#>  -1.3268 -3.4184  1.4571 -2.0524 -0.6424
#>  -2.5284  0.8367  6.1945 -0.6503  2.7010
#>   2.0027  4.9288  5.6596  1.3067  0.0239
#>   1.4044 -0.4670  8.0270 -1.0176 -1.2114
#> 
#> (1,4,.,.) = 
#>   -1.1573   6.1590  -4.5889 -12.9953  -0.2121
#>   -8.4341  -3.1781   7.6065   8.1505  -1.6865
#>    4.2254   2.1280  -6.6039  -2.3800   3.3437
#>    3.9639  -3.4904  -1.0005   3.3881   2.8390
#>    1.1847  12.3460  -9.4090  -2.2205  -1.3072
#> 
#> (1,5,.,.) = 
#>    4.5149  -0.7203  -5.4425   5.1364   2.4475
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]