Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    3.3921  -2.0301  -3.1022   3.9020   2.1640
#>    0.5272  -1.2125   6.8788   8.6678  -6.4663
#>   -1.2035   2.8136   8.0089   4.1795  -4.7377
#>    6.8111 -10.0699  -4.8536  -9.8888  -4.3988
#>   -8.7011   7.9959  10.4442   3.8099   5.5787
#> 
#> (1,2,.,.) = 
#>   2.2738  1.5716  2.7758  2.6105 -6.7666
#>   3.2002  4.6693 -1.2361 -2.1859  2.4074
#>  -1.2278 -2.9927  0.0886 -1.5858  1.8130
#>   3.9273  1.8041 -0.1948 -5.5681 -8.4092
#>   0.8246 -4.2451 -4.8201 -0.6903  0.1390
#> 
#> (1,3,.,.) = 
#>   -4.8278  -0.9230  -4.2918   5.3793   6.7869
#>   -3.0011   9.8914   0.3345  -0.6256  -7.1801
#>   -5.3663  -3.8119   1.6088  -3.2124   3.5260
#>    4.4565   7.3504   4.4044  -2.6532   0.1269
#>   -5.6059   3.9106   5.5030  11.0179   5.8242
#> 
#> (1,4,.,.) = 
#>   -2.2029   6.9443   0.7952  -0.9655   9.6027
#>    5.7562  -3.0980   1.6107   1.2706  -4.4434
#>   -1.1467  -1.4235   2.1576   1.6718   1.6275
#>    7.8868   4.7374  -6.7218   3.6105  -4.8519
#>   -2.8023  -9.1258   4.9102   5.0801  11.3285
#> 
#> (1,5,.,.) = 
#>   3.4519  4.8882  4.0028  2.5179  7.7968
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]