Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   2.6520 -1.3369 -3.9107  1.5269 -1.4217
#>  -7.2314  1.4500  5.9957 -6.4536  6.5244
#>   2.4454  3.8891 -6.1646  8.9916  2.3445
#>  -6.4429 -1.8468 -1.1161 -7.3855 -4.0961
#>   0.5167 -0.5646  2.4952 -1.2042 -3.5837
#> 
#> (1,2,.,.) = 
#>    6.1905  -1.3177   6.5627   1.7494  -0.8616
#>   -4.5565   6.5509   2.7820  -2.8829   3.5063
#>    9.9967   5.9750   6.5483  16.8025   0.6982
#>    0.8649  -8.8551   0.2290  -6.4223   5.0706
#>    0.6187   8.3542  -0.8636   5.7229   0.0854
#> 
#> (1,3,.,.) = 
#>    0.8872   0.9206  -5.7294   3.7029   3.9039
#>   -0.4843   0.7334  -9.6267  -4.1496 -10.6114
#>    7.9304   2.5315   5.7397  -1.2297   0.8944
#>    3.6792  11.6725   4.6402  -4.4534   2.0463
#>   -3.9542   7.0342  -2.3756  12.4159   1.1558
#> 
#> (1,4,.,.) = 
#>  -0.3077 -3.4047 -3.0652 -6.3187 -0.8986
#>  -1.9464 -2.8125 -2.1636  2.9646 -4.7152
#>  -1.5828 -2.3306 -4.2561 -1.7756 -0.3142
#>  -4.3594  2.1155 -4.6234 -3.9137 -9.1309
#>  -3.5490  2.7592  0.5698 -5.5326  0.0436
#> 
#> (1,5,.,.) = 
#>   -0.9072   3.0823   3.4394   1.9713  -0.1417
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]