Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   0.6388 -5.8630 -4.0296  0.3027  2.8360
#>   3.1054  4.4387 -6.1227  6.5574 -1.7879
#>  -2.6164 -2.4727  0.4637 -0.1077 -13.0619
#>  -4.0760  1.6378  4.8474 -3.2228 -9.7162
#>   2.8823  9.2779  6.0286  4.3111  3.1992
#> 
#> (1,2,.,.) = 
#>   2.6507 -1.7286  7.3980 -4.6938  0.3164
#>  -2.8122  4.3606 -5.6012 -6.1309  2.4105
#>   5.6314  7.2271 -6.3551  6.5785  0.7611
#>  -0.0274 -5.0980  0.8699 -1.7682  0.8320
#>  -1.8254 -10.7792  1.6284  6.2144  0.4067
#> 
#> (1,3,.,.) = 
#>   -4.3118  -0.7798  -9.1615  -0.6679  -2.0034
#>    3.4604  -0.2423  -9.0536  15.0302  -9.0469
#>   -1.9987  -8.3936  14.6244  13.6022  -1.5092
#>    1.6906  -0.7235   3.0303  10.4590   1.7651
#>    2.0904  -5.9218  -8.5562  -6.5026   0.3127
#> 
#> (1,4,.,.) = 
#>    0.1872  -0.2283   3.5690  -7.5915   2.2861
#>   -4.9479  -1.8985  -2.7869  -1.7106  -0.0248
#>   -3.8708  14.5540   4.6667   2.8116  -2.0290
#>   -2.8647   0.9203  -1.7870  -6.2141  -0.8053
#>    3.1509   5.5500   7.9885   1.4815   0.9182
#> 
#> (1,5,.,.) = 
#>   -1.0276  -3.0941   7.2655  -0.4251  -7.3026
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]