Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   2.1750  3.4405 -0.8077 -0.8297 -1.5692
#>   0.0861 -4.3701 -0.6045 -2.2217 -8.3044
#>   6.1283 -0.7893  5.5514 -0.8434  4.6148
#>  -6.5673 -8.4730 -0.0850 -5.0130 -5.4617
#>   1.8666  1.4001  4.2978  3.5239  4.1132
#> 
#> (1,2,.,.) = 
#>   0.1850  5.2744 -1.8433  2.2437 -1.9356
#>   3.0377  2.8512 -9.1646  6.7076  3.9060
#>   1.7996 -1.3667 -4.5409  3.2772  2.5190
#>  -3.3464 -5.0462  0.8031 -2.4342 -4.9948
#>  -2.6480 -1.4681  1.9025 -3.1124  3.5450
#> 
#> (1,3,.,.) = 
#>    0.9763   1.7468  11.9755 -11.1550  -0.0937
#>    2.8089  -8.7713   3.6391  14.5444  -3.1190
#>    4.4255  -0.8481   8.8436   3.7328  -2.7583
#>   -8.5816  -2.3860  -6.9171  -3.5598   0.1871
#>    2.7786   8.0085   7.3156   6.5143   0.8353
#> 
#> (1,4,.,.) = 
#>    5.2159   6.7596   4.8569 -11.3406  -5.1906
#>   -2.9887   1.6025   8.4872  -0.9977   4.2509
#>   -2.5004   0.1751   8.8407   6.5655  11.9557
#>   -8.0732  -1.3077  -3.5579  -4.8650  -3.6981
#>   -1.8625   8.7038   3.3273  18.9684   2.6568
#> 
#> (1,5,.,.) = 
#>   -7.1857  -5.6273  -4.9066   5.0509   2.7254
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]