Conv_transpose2d

Conv_transpose2d

Arguments

input
NA input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
NA filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
NA optional bias of shape \((\mbox{out\_channels})\). Default: None

stride
NA the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
NA dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
NA additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
NA split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
NA the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=None, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See ~torch.nn.ConvTranspose2d for details and output shape.

.. include:: cudnn_deterministic.rst

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -7.5962 -3.3326  6.7831 -2.6847 -3.8770
#>  -3.7980 -14.6025  5.4740  7.9556  2.2569
#>   1.0895 -3.0573 -4.4158 -2.3707  6.6306
#>  -4.1577  8.8056  8.9085 -7.4639 -3.1753
#>  -4.5997 -5.8294  7.1765  4.1298  1.2998
#> 
#> (1,2,.,.) = 
#>  -1.2496  1.3804  0.3830  2.8516  1.8649
#>  -0.9480  9.5710  3.6643  5.1610 -2.5484
#>   1.5285  2.5909 -4.4857 -1.0627  5.6590
#>   2.2464  6.2198  0.5224  1.8569 -7.0381
#>  -2.6082 -0.3835 -8.3964  1.6590  1.8936
#> 
#> (1,3,.,.) = 
#>   -5.4240  -1.8959   2.0982 -11.0572  -7.1129
#>   -5.7286   8.2632  -1.4871  -9.0047 -12.1764
#>    0.1466  -8.4095   1.8028  10.7652   7.6637
#>   12.9905   2.8676  -1.1665  -6.1606  -7.3444
#>    5.2231  -2.5144  -5.5655  -5.1334  -1.7502
#> 
#> (1,4,.,.) = 
#>  -3.1346  9.0839  4.9047 -5.5578 -1.4678
#>  -2.2756  0.3479  8.9028  4.4772 -3.2621
#>  -5.0682  1.0232  3.3966  2.8198  1.8310
#>  -7.1607  5.5815 -6.9231 -7.6175  1.6746
#>  -11.2850  2.7781  2.8169  7.0857 -6.2299
#> 
#> (1,5,.,.) = 
#>   -6.6460  -4.2459   1.5838   2.3837  -0.8733
#>    2.7812  -9.1419  12.0371   9.8969  -0.8947
#>    0.7636  -5.1756   5.2435  16.0599   2.2810
#>   -0.8439  -2.0671  -7.5943  -0.5896  -6.8856
#>    2.4704  -8.5656   4.3656   0.3382   0.3326
#> 
#> (1,6,.,.) = 
#>    4.7029  -3.6134   1.1732   5.8723  -0.2575
#>    3.7304   1.1764  -1.7565   8.2722  -2.0145
#>   -5.3253  -3.6576 -12.0914   3.9391  -0.8159
#>    5.3560   2.9393  11.4436  -3.2281  -9.4492
#>    0.9640  -6.3274  -7.7627   5.7794   4.3723
#> 
#> (1,7,.,.) = 
#>    1.0651   5.3157  -2.6739  -2.2390   3.0830
#>    3.6786  13.4648   6.7157  -5.3990   0.1457
#>    4.9947  -4.2912  -6.5892  -1.3185  13.0362
#>   -4.1518   3.5200   5.8145   7.6304   4.5048
#>   -3.7006  -5.7260   4.4005  -6.7617   5.6349
#> 
#> (1,8,.,.) = 
#>    0.2253   4.1118   3.4038   0.7874  -1.8600
#>    1.7147 -10.3997  -5.7130  -5.1718  -0.9915
#>    2.5338  -7.5954  -9.8617  -5.6232  -4.2832
#>   -6.7388  -1.7717   0.5150   2.3745   0.4394
#>   -4.0281   2.9175  11.6638   6.1722  -0.0937
#> [ CPUFloatType{1,8,5,5} ]