Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    7.5408   6.3594   5.6068  -0.3825  -2.7078
#>   -4.7164   5.8063   3.3500   3.6371  10.3153
#>   -4.2331   8.8011   5.5059  -7.5571   1.4603
#>   12.5013  -2.6130   4.3408  -6.0741   6.8690
#>   -0.3311   3.2887  -3.0918  -3.3116  -4.5239
#> 
#> (1,2,.,.) = 
#>  -4.6556 -5.9922 -3.9867  1.8707  6.7915
#>  -2.1588  2.7291  4.5598 -2.0503  4.1656
#>   6.2416 -6.1598 -17.0948  2.6983  1.0388
#>  -5.0301 -1.5794 -0.6743  2.2365  2.8580
#>   7.5425  6.4918 -0.1441 -2.7134  2.2018
#> 
#> (1,3,.,.) = 
#>  -1.3987 -0.7297 -0.2402 -3.7806  3.0967
#>   6.1404 -3.3849  7.3881  3.9147  5.0814
#>   1.8310  3.5506  2.6034  0.6772 -1.4086
#>  -2.2774  0.0658 -3.7855  0.1120  0.3979
#>   0.1161 -2.0250  2.7839 -12.5085 -1.6722
#> 
#> (1,4,.,.) = 
#>   -4.3826   5.6915  -4.4417  -2.3415   3.2487
#>  -10.4974  11.3892  -5.5171  -7.5938   8.7782
#>    2.2063   2.3368   1.6665  -4.9180   5.4083
#>   -5.9692   3.1180  -1.0278   2.0040   5.4752
#>   -3.8039  -1.2763  -8.1951  -0.2508  -3.1128
#> 
#> (1,5,.,.) = 
#>   -4.6516  -0.6603  -2.7194   3.9354   0.8557
#>    1.9038   4.6945   1.6821  -5.3679  -1.9047
#>   -2.6784   2.6000  -9.7269  10.4007  -2.5007
#>   -1.0439  -7.8244  11.0364   9.7092  -6.3770
#>    0.9664   2.4206 -12.9538   2.1692  -3.5030
#> 
#> (1,6,.,.) = 
#>   -4.1428  -1.4510   6.8454   6.5473  -1.5271
#>   14.7920  -6.2547  -2.9182  11.3248   3.7863
#>    3.5591  10.0227  -8.4269   3.3137   7.6180
#>    2.1815   2.6556  -7.4308   3.5433  -0.5957
#>    0.1434  -5.0020   5.8209   4.7412  -1.6124
#> 
#> (1,7,.,.) = 
#>   2.2678  5.7145 -2.8558 -0.8086  4.7832
#>  -2.1524  1.9802  3.2322  0.5380  0.7971
#>   2.6210  2.8738  0.7705  2.9230 -6.2050
#>   1.4428 -11.7758  4.6113  2.8551 -7.1939
#>   2.2852  4.7497  7.1903 -2.2664 -6.4840
#> 
#> (1,8,.,.) = 
#>  -3.9711  4.8101  7.8477  2.1330  1.6587
#>   0.8643 -4.7801  7.8337  0.6497  5.0487
#>  -4.9889 -2.2494 -4.3536 -0.7439  1.1247
#>  -0.8267  0.6191 -9.2026 -2.3405  8.6275
#>  -2.2285 -6.4247 -4.9930 -5.1804 -1.1252
#> [ CPUFloatType{1,8,5,5} ]