Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   5.2264  7.1965  2.3252  1.8624  8.5392
#>  -3.1061  9.9300  3.4257 -0.7107  6.5695
#>  -7.4311 -3.0500  1.9592 -5.8037  0.9992
#>  -1.7743 -0.1771  4.7801 -8.6383 -8.4588
#>  -1.9371  2.5641 -0.8673  4.0312 -6.0665
#> 
#> (1,2,.,.) = 
#>   -1.8744  -4.9983  -2.6170  -9.0904   3.0888
#>    1.9278  -1.1812   1.8911   0.7460  -4.7496
#>    3.7994  -6.0113  -0.8291   4.8027  -1.4948
#>   11.0966  -8.4557   3.2276  10.1932  -4.3066
#>    6.1695  -6.7672  -5.0733   1.9290  -0.5934
#> 
#> (1,3,.,.) = 
#>  -1.6631 -3.0472 -1.2148 -1.7890  8.8696
#>   2.6545 -10.0469 -0.7205 -4.0600 -0.6493
#>  -1.3020  3.8496  3.8330  1.4661 -1.2852
#>   0.0271  0.1491 -1.8297  5.5546 -2.5041
#>  -0.8909  2.7930  1.5010 -3.7005 -4.3194
#> 
#> (1,4,.,.) = 
#>  -4.5225 -3.7189 -5.6428  2.0973  2.4071
#>   2.3477  9.0525 -0.6345 -7.6951  1.2453
#>  -8.8588 -9.9055 -4.7376 -0.9127  3.6565
#>   5.0937  0.1332  5.2492 -0.3656 -0.7763
#>   2.5453 -8.6117 -3.0789  7.9309  1.4514
#> 
#> (1,5,.,.) = 
#>   -1.1785   5.3104   2.1863   1.0626   1.4794
#>    3.7795  -8.9725 -11.2851  -2.4326  -4.5428
#>    4.7288  -3.1366  12.9042   2.5537  -9.4836
#>   -1.3011   2.2748   4.1159   6.8862  -0.4677
#>   -3.2113   4.0786   0.0906 -10.1024  -5.0755
#> 
#> (1,6,.,.) = 
#>   -1.7933  -4.6473  -0.6231  -2.9634  -5.8635
#>   -0.2984  -1.1307  -2.4370  -7.5725  -2.4641
#>    1.7340   3.9035  12.9656   0.7024   1.6121
#>    0.7374  -2.7035  -1.2514  -7.5804   0.8820
#>    0.0647   2.4143   1.2932   1.7039   1.3565
#> 
#> (1,7,.,.) = 
#>    0.7963  -2.0029  -3.3805  -0.6467  -3.3152
#>   -0.2312  -5.3789  -2.1990  11.0259  -3.1839
#>    2.7080   1.0884   1.1216   2.5039  -2.2630
#>    2.6911   1.9701  -5.2696  -1.1672   1.5094
#>    1.4002   3.3327  -0.4173  -3.0324   8.4496
#> 
#> (1,8,.,.) = 
#>   1.4346 -4.2876  1.1120 -9.6294 -4.0842
#>   3.9162 -1.7977  9.2921  1.6556 -1.0478
#>   5.2038 -3.1772  5.0023  0.6024  2.4384
#>  -0.0855  1.7869 -0.4271  3.2897 -5.9415
#>  -1.8278  6.5046 -2.9712 -1.2234  0.0828
#> [ CPUFloatType{1,8,5,5} ]