Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    8.2926   1.9182  -4.5741  -1.3766   0.5506
#>   -7.2540   3.6096   2.3139   5.4672  -0.9078
#>    9.1255  -1.9232  11.5756  -5.3749   0.2204
#>    3.1942  -2.7864  -0.1492 -19.1416   2.0583
#>    3.3746   0.0500  -3.0366   2.7356  -4.7539
#> 
#> (1,2,.,.) = 
#>   0.9671 -6.5849 -6.4200 -2.9053  2.3344
#>   9.1160  0.0355  4.5259  6.7737  2.2024
#>  -7.8460 -6.5574 -9.8855  1.5913 -7.7886
#>   5.7660  5.3995 -4.7880  8.0046 -0.5061
#>  -0.7804  3.3259  3.4393 -9.7218  1.0270
#> 
#> (1,3,.,.) = 
#>   -2.4384   7.3718   1.3365   1.1912  -5.2541
#>    1.9014  -0.8376  -4.0008  -6.2616   7.2093
#>    1.3209  -9.0596   5.1543   4.8816   0.7328
#>    2.8906  10.1036  -2.9487   1.6142  -3.9229
#>    0.1221   1.8300   1.3589  -7.6201   1.7108
#> 
#> (1,4,.,.) = 
#>   -5.5476  -0.1702  11.2181  -0.3949   3.5693
#>   -1.8751   9.5813 -11.8790 -13.7090   1.8872
#>    3.9674   0.2641   1.2384  15.0161   5.6357
#>    0.3652  -1.5320  -5.9489  -0.9193   8.8908
#>   -2.7543  -1.2582   7.3219   1.5477  -8.8204
#> 
#> (1,5,.,.) = 
#>  -8.1173 -0.5053  5.8105 -2.1245 -3.7749
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]