Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   3.2829  5.4719 -0.5863  1.4491  6.4350
#>   4.0812  2.1737  5.2615  1.3419  8.5702
#>  -2.2863  4.9520  4.8247 -14.2278  5.6260
#>   5.9185  4.2258 -6.5793 -13.5251  1.5357
#>   0.3379 -4.5384 -1.2461 -1.8918  6.1181
#> 
#> (1,2,.,.) = 
#>  -4.6999  3.0426  5.2413 -9.7982  0.5961
#>   5.4036 -3.8291 -10.7679 -6.1173 -7.6761
#>   5.0196 -3.6155  7.8543  4.2008  3.8775
#>  -9.4034  2.8434  0.1118  4.4905 -1.8465
#>   1.3589  1.9051 -2.4784 -0.7444  0.9538
#> 
#> (1,3,.,.) = 
#>    4.7865   2.6940  -5.3823   2.0613  -0.5751
#>    1.8424   6.4023   7.3333   2.0236  10.0820
#>   -3.4784   1.0719   4.1365 -10.4323  -4.1934
#>    7.4523   3.7278  -0.9012 -13.0564  -2.3261
#>   -0.4172  -4.0714  -3.8005  -2.4769  -5.5445
#> 
#> (1,4,.,.) = 
#>  -4.0850  4.1374 -6.3191  2.7402  3.2905
#>  -3.0997 -1.0082 -4.5584  1.0341 -1.1907
#>   1.0798 -1.6023  5.9377 -1.5026  3.9412
#>   0.9470  0.3319  0.9509 -7.2008 -6.3662
#>   4.9034  4.4610  4.0362 -3.7771 -3.3909
#> 
#> (1,5,.,.) = 
#>   1.9219 -0.6470 -2.5047 -1.0283 -4.8451
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]