Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -3.1987 -2.2647 -2.0334  0.9186  2.0237
#>   7.9880  4.7540 -9.3711 -6.0775  5.7953
#>  -4.7061 -1.1319  4.3198  3.3864  2.4688
#>  -0.5580 -2.6302 -2.7481  1.7361  9.3793
#>   5.7431 -0.1393 -4.8614 -2.5453 -7.1517
#> 
#> (1,2,.,.) = 
#>   -0.8664   7.9211  -3.0997  -5.8358  -0.5269
#>   -4.2749   7.3147  -7.5922  -8.3301  13.1288
#>    1.9189  -2.0340  -5.5973  -3.0048   4.6235
#>   -1.5038   6.1146 -10.6041   5.4299   5.1156
#>    4.3177  -0.5559  -1.5447   3.0581   0.1938
#> 
#> (1,3,.,.) = 
#>  -0.7517 -0.4835  0.4102 -0.9885  1.1577
#>  -7.6656  4.0259  1.6894  0.8315 -4.9959
#>  -1.8840  0.6775  2.2833 -3.8618 -6.4874
#>  -6.1756 -4.2269  1.5882 -5.1988 -0.0211
#>  -7.5088  5.1715 -3.6360  3.6597 -1.9512
#> 
#> (1,4,.,.) = 
#>   1.6115 -6.9024 -0.7302 -0.3584 -1.1757
#>   0.7375 -5.7653  1.6608  7.7836  0.4787
#>   5.4374 -0.0091  1.6512  1.1040 -3.3797
#>  -0.4900 -0.8811 -0.7764  9.9311 -5.3977
#>  -3.7801 -0.5193 -5.4596  3.6966 -4.1438
#> 
#> (1,5,.,.) = 
#>   2.9961  0.3427 -2.2493 -0.3616 -4.0308
#>   0.1140 -0.2408  5.6699 -3.0243  9.2990
#>   4.4366 -5.5484  2.6653 -3.2188  0.8404
#>  -1.3681 -0.6632  1.0805 -5.1564  2.6230
#>   1.0123  2.1613  0.2220 -2.1670 -4.4421
#> 
#> (1,6,.,.) = 
#>   -4.4281  -1.5713  -1.6736   2.2781  -3.9112
#>   -3.2605   8.2142 -10.9234  -8.8717   9.4848
#>    2.9832  -4.3320   1.6529  -4.6635  -1.1412
#>   -7.7500   3.9805 -13.2990  10.6255   3.7598
#>    2.8888   1.8326 -10.4985   8.0102  -7.5440
#> 
#> (1,7,.,.) = 
#>  -5.9033  5.3286 -0.2104 -5.2385 -5.7057
#>   0.1725 -7.4607  8.6631  4.4932  5.8086
#>   5.7449  0.2322  1.2773 -12.0418 -2.6350
#>  -4.8518  5.8266  4.2042  3.5178  1.1046
#>   0.4453 -5.5504  4.6868 -8.3188  2.6615
#> 
#> (1,8,.,.) = 
#>   0.4494  2.2916 -14.0544  2.0116  5.1031
#>  -2.9357 -0.2364  5.4885  3.4106 -0.8578
#>  -0.7084 -0.2138 -5.5943  3.7190  5.1977
#>  -2.1151  3.5114 -0.2189  6.2124 -7.5526
#>  -1.8996 -3.5934 -1.2281 -1.5988 -0.9148
#> [ CPUFloatType{1,8,5,5} ]