Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   0.4502  5.8268  5.0776 -8.1647 -3.3241
#>   1.0104  1.8673 -2.7223  5.9275  2.5404
#>   4.2399 -6.8440 -3.4329 -4.8705  2.6759
#>   1.2954  6.9710  0.5624 -5.5562 -2.6579
#>  -1.9093  3.2094 -2.0327  0.1294  3.2344
#> 
#> (1,2,.,.) = 
#>   -1.6026  -1.5015  -4.3140   2.3352   0.1920
#>   -6.4574  -0.0677  -8.1785   7.1825  -4.8838
#>    0.5326   5.6479  -0.5504  -4.3836  -6.2505
#>    1.5188   7.4418  10.0081  -2.8776  -5.8677
#>   -2.2919   1.6159  -0.4096   2.6166   4.1143
#> 
#> (1,3,.,.) = 
#>   10.2432  -7.1416  -3.8276   1.5012   4.9983
#>    0.4155  -5.4071  -1.6285  -5.5137  -0.3016
#>    0.1665  -1.2446  -7.2279   2.3814   3.3515
#>   -0.3505  -3.6859  -4.0521  -3.2525  -1.3862
#>    2.8365  -0.9474   1.3336  -7.5780  -2.0589
#> 
#> (1,4,.,.) = 
#>  -2.4079  7.2058  9.1735  3.0084 -1.2767
#>   7.9296 -4.2180 -14.4949 -1.4559  0.2371
#>  -4.9178  0.0492  8.2585 -5.7345 -4.4876
#>  -1.0942  1.2482  0.8184 -0.9157 -2.1248
#>   5.6164  5.2152  3.4339  4.1216 -2.0079
#> 
#> (1,5,.,.) = 
#>   -7.1519   5.2091   1.0564   1.0655  -4.3576
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]