Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -2.6433  3.7040 -0.2507  5.5767  0.2008
#>   2.1883 -8.7574  3.7210 -2.3549  9.2242
#>  -0.2680 -0.8045  3.1260  6.3419 -3.1812
#>  -1.2699 -3.2510 -2.1596  1.8070 -5.9565
#>  -7.8361  4.9789 -0.7123 -3.7312  3.8328
#> 
#> (1,2,.,.) = 
#>   0.8764 -5.3936  1.8305 -0.5427  6.5909
#>  -1.9885 -5.3874 -7.8525  2.1417  1.5280
#>   0.2950  5.6897 -3.2537 -0.4161 -3.1010
#>   0.4234  5.9786 -7.0275 -1.8250 -7.9552
#>  -6.5597 -6.7512  2.7030  5.9729  3.5624
#> 
#> (1,3,.,.) = 
#>  -0.2708  2.0922 -13.1177  1.5573 -6.3868
#>   2.8845 -0.6535 -9.7343 -0.4852  0.0742
#>  -0.0367 -7.7990  5.9951 -15.4424  3.6144
#>   7.0965 -13.7556  0.5675  8.5965 -2.5498
#>  -7.0791 -1.1796  2.2402 -3.0806  3.3600
#> 
#> (1,4,.,.) = 
#>   0.1914  2.9437  0.1111  3.5610  4.7904
#>   2.7272 -2.6153  0.2802 -4.1900 -1.3629
#>   4.6449  0.5301  0.3462  4.0408 -5.4633
#>   2.8259  3.0289  4.7003 -0.4462  1.5977
#>  -0.6193 -0.6820 -0.0065 -8.2069  2.2825
#> 
#> (1,5,.,.) = 
#>    0.3761   5.0140  -4.0007  12.6900   0.3758
#>   -2.2375  -4.2463  -4.6936   2.9703   0.1946
#>    2.9191   2.8709   6.1963   1.7056  -2.6945
#>    4.8172  -8.7045  -0.1965 -11.4245  -1.7360
#>   -1.6763  -6.8268   0.5968  -3.8070   6.6505
#> 
#> (1,6,.,.) = 
#>  -2.5610 -0.3681  8.5378 -7.9547  1.7665
#>  -2.5091  4.5104 -4.8141  4.1874  2.9702
#>  -0.3462  6.0468 -5.5500 -5.9201  3.7018
#>  -6.9249 -6.5752 -6.3991  0.0123  4.2602
#>   0.9347  5.5996  0.6497  6.6477 -4.4275
#> 
#> (1,7,.,.) = 
#>  -3.3505 -0.1735  5.1380 -8.1403  3.7246
#>  -4.5435  9.1582 -3.3728  0.1209 -0.6642
#>  -0.7822  4.5197 -7.2627  4.7848  4.2291
#>  -1.4667 -3.6035  8.9916 -1.9156  3.3885
#>   1.9070 -5.2830  1.9904  2.6850 -3.3405
#> 
#> (1,8,.,.) = 
#>   2.1791 -6.5561 -1.0909 -2.7120  0.7352
#>   2.4313  1.6129 -0.4255  7.9818  2.8188
#>   1.3777 -5.0150  1.8077 -1.3884  3.4954
#>  -4.8596 -3.5585  2.9971  2.3842  6.0421
#>  -0.6497  0.5285 -3.8733  3.8083 -5.3335
#> [ CPUFloatType{1,8,5,5} ]