Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    3.9140  -6.7131   4.3647  -9.0615   5.6857
#>    0.7231   0.5531  -4.1240   8.1279  -1.6242
#>   -3.7923 -18.1069   3.4200   0.1742   0.3456
#>   -2.9511  -5.1799  -4.6907   6.2121 -10.3887
#>    0.6942   1.9005   4.3800  13.9836   0.6363
#> 
#> (1,2,.,.) = 
#>    3.3144  -1.8912  -5.1244   5.3264   0.5261
#>    3.9226  -2.4262  12.2498   5.4956  -4.1487
#>    0.9134  -5.8987  -0.3195  -1.0592  -5.3113
#>    6.3188  14.0448 -17.1054  -0.3309  15.7261
#>    4.6686   0.8861  -2.0497   0.3060   0.5930
#> 
#> (1,3,.,.) = 
#>  -0.1587 -6.5169  6.4862 -0.3807  0.2753
#>  -0.8699 -0.2326  1.3061 -1.6784 -8.9877
#>   2.9172 -0.7881  9.6576  2.4117  0.7062
#>   1.6744  1.2269 -7.8282 -0.1233  4.3300
#>   2.2978 -0.8797 -0.5959 -4.2996 -2.3489
#> 
#> (1,4,.,.) = 
#>    1.9808  -4.8039  -0.0726   1.0135  -4.8478
#>    3.1675  -4.6491  -6.3618   1.5396   1.0784
#>    2.7723   3.6929  -8.2700  16.0420  -5.9048
#>   -5.0766   0.7083   7.6275   0.6653   6.8291
#>    1.3760   9.3372   5.4603   1.9790   9.2604
#> 
#> (1,5,.,.) = 
#>    0.1746  -0.6704  11.1931  -5.3597   4.1758
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]