Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -3.3648  -1.1255  -0.3963   4.6891   1.6278
#>    0.4268   7.1008   3.3606  10.1274   0.7294
#>    4.4011  -9.0175   6.1489  -3.0525   8.6341
#>   -6.7353  -4.8900  -0.0131   1.4166   1.9201
#>   -2.1622  -3.2308   3.1979  -0.4882  -4.1013
#> 
#> (1,2,.,.) = 
#>  -2.2240  6.1296 -1.0868 -2.8141 -2.6927
#>  -0.0064 -7.3219  3.6651 -6.7182 -1.2233
#>   4.7290 -2.4576 -0.3048 -0.9654  7.3283
#>   1.8467 -5.9483  2.2162 -0.2004  7.5238
#>  -2.6903 -2.5052  2.8450 -1.2584  1.8917
#> 
#> (1,3,.,.) = 
#>   -0.2429   7.9687  -7.0417  -4.2401  -1.6047
#>   -4.5318   0.5429   4.0789  -1.9303  -3.9097
#>   -1.7556  -0.0897   6.9972   5.2794   2.0335
#>    2.0999  -2.2292  -4.9530  -1.0641   1.4518
#>    2.3684   7.1330   1.7770  11.6271   5.9480
#> 
#> (1,4,.,.) = 
#>  -1.3450  1.1131 -6.4920  1.2516 -1.7022
#>   3.5330 -2.9936  2.5531 -7.2964 -0.4000
#>   1.5349 -6.9951  9.7438 -0.2257 -2.5552
#>  -3.8505 -0.0733 -10.7112  5.2209 -9.7751
#>   6.4369  1.4316 -6.8895 -0.7854  6.0959
#> 
#> (1,5,.,.) = 
#>   1.0312  0.4831  0.1458  2.4165 -6.8746
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]