Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -1.2753  4.6754  0.2125 -8.1016  0.8186
#>  -4.1125 -8.7817  6.8819  7.4590  3.6260
#>   0.5207 -8.7950  2.1031  0.3167 -2.5377
#>  -2.1374  0.8132 -1.2757 -1.1917 -5.8558
#>  -2.0681 -0.3958  5.1496 -2.6883 -0.4483
#> 
#> (1,2,.,.) = 
#>  -2.5371  1.3020 -1.4070  0.7440  8.2851
#>  -6.2105 -7.4518  2.8559 -5.7133 -9.6730
#>   9.5100 -3.2850  0.9348  3.4341  2.0955
#>  -4.2809 -13.0464 -4.6652  6.9445 -4.5938
#>   6.1264 -0.5239 -0.7266 -2.6405 -1.0276
#> 
#> (1,3,.,.) = 
#>   2.3701 -0.2923  0.0796 -1.0214  0.7015
#>  -0.7239  7.4763  5.5189 -5.1072 -3.9999
#>   0.3463  0.5533  1.6939 -1.2899  1.2758
#>  -5.6493  2.1973  1.9242  2.2693  6.0706
#>   3.1029 -13.2316  2.8093  1.1034 -4.8092
#> 
#> (1,4,.,.) = 
#>    4.0628  -4.1283  -7.2233   2.6129   0.9194
#>   10.3867   1.8575   0.4421  -0.3779  -0.2188
#>    1.6037  -3.7622   3.2286   1.8191   1.3954
#>    6.7991   1.7581  -3.7914  -4.6854   4.7185
#>    9.1870   0.2211   1.7581   0.2606   2.2811
#> 
#> (1,5,.,.) = 
#>    0.5767  -6.6006   1.2972  10.0892  -5.7840
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]