Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    2.2788  -1.9283   7.3256   5.1428  -2.9257
#>   -0.6896   1.4029  -5.3229  10.2390  -1.2402
#>   -1.2060  10.2222  -4.7248  -1.5488   3.8320
#>   -1.4776   1.1281   1.8286  -4.8006   8.1204
#>    5.9955   0.2092   0.9957  -2.9968  -1.0158
#> 
#> (1,2,.,.) = 
#>  -4.0071 -2.0428 -1.4256 -1.4038 -0.1210
#>  -5.2884 -1.0475  7.1233  1.1916  5.0784
#>   3.7178 -1.1443  2.2506 -3.7226  0.3739
#>  -4.2545  2.0034  2.7678 -2.2835 -0.9902
#>  -3.2392  3.3695  1.1751 -1.6468 -2.7825
#> 
#> (1,3,.,.) = 
#>   0.5532 -7.5668  2.3785  3.4235  2.2801
#>  -0.4576  0.1908  3.4579  6.5030  2.8629
#>   5.5145  5.7911 -5.0308 -4.4961  0.1457
#>   0.1480 -0.5277 -3.9319  6.4784 -1.5596
#>  -0.1217  0.2231 -2.0216  2.9053 -0.2022
#> 
#> (1,4,.,.) = 
#>   0.3142 -8.9630 -1.8311 -2.6733 -1.9425
#>  -6.8940  1.7532 -5.0590 -4.0202 -4.1672
#>   5.6343 -3.8787 -7.2058 -10.6780  1.3036
#>  -1.8006 -1.8730 -5.1896  5.3562 -0.2618
#>  -5.8934  0.0295  0.0948  1.0247 -2.2523
#> 
#> (1,5,.,.) = 
#>   5.1338 -0.8517  5.0228  3.4769  3.7733
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]