Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -0.3143 -2.9301  7.5563  0.6644  1.0507
#>   1.1478  1.1548 -7.4076 -2.7759  3.1744
#>   3.0819  6.7321  2.0493  0.5808 -2.4675
#>   4.9198 -4.9058 -0.0826 -11.0084 -5.7684
#>   1.6682 -2.6875  3.4297  2.5959  4.1683
#> 
#> (1,2,.,.) = 
#>   0.1772 -0.2208 -3.0126 -6.9597 -4.4578
#>  -3.6811  1.2047 -0.0595 -4.4851 -2.1285
#>   4.8564  3.9823  5.3009 -0.7482 -3.1789
#>   1.8347  0.4092 -4.2792 -3.9713 -6.2866
#>   2.5404  2.9659  0.7068 -4.2953  1.8262
#> 
#> (1,3,.,.) = 
#>   1.0805 -2.7649  0.4549 -1.8160 -0.7879
#>  -5.6020  2.3493 -6.5694 -3.4869  3.9372
#>   7.2545 -2.9695  4.6283 -4.0261  0.0073
#>   3.5666  0.9623  0.5716  0.5572  3.2985
#>   1.3883  0.0650  1.1537 -0.1674  1.4602
#> 
#> (1,4,.,.) = 
#>   0.0621  1.7024  2.2315  1.1524  2.1682
#>  -0.3297  0.4764 -7.6606  0.1708  2.6889
#>  -0.9199 -2.5920 -1.4577  0.3050  4.0339
#>  -0.2253 -2.4640 -2.0471 -1.5677  1.4351
#>  -5.2765  0.0563  0.2994  3.7422  2.5269
#> 
#> (1,5,.,.) = 
#>   1.5132 -0.3808  1.4774 -5.1841 -7.3635
#>  -2.8323  1.7472 -3.6173 -2.3114 -1.9975
#>   3.9249  7.0658  4.5332 -3.4794  3.9799
#>   8.2268 -2.7744 -6.8575 -2.9755 -4.4241
#>  -0.7401 -6.6266  0.2722 -6.1297  8.0833
#> 
#> (1,6,.,.) = 
#>   3.7256e+00 -4.1994e+00 -5.2859e+00  4.0028e+00  2.9519e+00
#>   5.0701e+00 -3.8331e+00  9.0122e-05  4.9578e+00 -4.8377e+00
#>  -2.9739e+00 -1.7368e+01 -7.6006e-01  8.9405e+00 -4.4929e+00
#>   4.5959e+00 -3.6691e+00  1.0554e+01 -5.8683e+00 -4.4625e+00
#>   1.9653e+00 -2.4343e+00  3.7473e+00 -1.7400e+00 -2.8814e-01
#> 
#> (1,7,.,.) = 
#>  -0.2546  5.0676  0.6538 -2.9942  2.1290
#>  -0.0275  1.0235  4.9340 -1.2646  3.5815
#>  -0.7316  4.5204 -2.4266  7.2611 -3.2295
#>  -0.2274 -1.8257  4.2993  8.4616 -1.7315
#>  -3.4160 -8.9939  6.8951 -7.2549 -4.1202
#> 
#> (1,8,.,.) = 
#>   -0.0092   2.7063   1.5104  -0.3622  -2.4561
#>   -3.9070  -1.0746  -1.3218  -5.6135   2.6430
#>    3.8127  12.4397   3.0715   1.9367   5.5469
#>    1.8837  -5.2359 -10.3395  -7.8063  -7.3797
#>    0.5561   6.0104   7.3829   2.4054   2.3034
#> [ CPUFloatType{1,8,5,5} ]