Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -1.0662e+00  2.0970e+00  1.8124e+00 -2.6491e-01 -2.4031e+00
#>   4.6217e+00  2.4802e+00  7.3807e-04 -1.2285e+00  2.4903e+00
#>  -3.2950e-01 -4.7274e+00  3.3830e+00  6.1694e+00  1.3272e+00
#>  -6.8243e-01 -3.3817e+00  1.3123e+01 -7.5971e+00 -7.4466e-01
#>  -3.3458e-01  8.6429e-01  1.9723e+00 -4.1259e+00 -1.1658e+00
#> 
#> (1,2,.,.) = 
#>   3.3719 -5.1180  3.1989  3.5474 -3.5832
#>  -2.3319  1.1619  4.0927 -0.1062 -7.2706
#>   3.2801 -9.4655 -2.5720  0.4160 -1.0844
#>   1.3459 -4.7464  8.8120  5.7952 -2.0792
#>   2.4652 -5.5812 -1.8681 -0.0671 -6.9541
#> 
#> (1,3,.,.) = 
#>   -5.3567  -1.6533  -2.0941 -14.2823   4.1224
#>    5.4431  -0.4583   3.5747  -0.3895   3.7259
#>    0.9781   0.0340  -9.3191  10.9623  -0.5972
#>   -2.9044   2.5198  13.6096  12.3245   2.9534
#>    4.5144   9.1297   3.4572   1.1362   2.6823
#> 
#> (1,4,.,.) = 
#>   -1.8811  10.5431   4.7537  -2.5476   3.4887
#>    4.9210  -4.3821   3.2979  -6.3451   3.2680
#>   12.6958  12.4486   1.2362  12.2827   0.7768
#>    3.1705   8.4931  -1.3204   1.1024  -1.5411
#>   -5.3649   2.3643  -5.8019   8.4187  -2.4314
#> 
#> (1,5,.,.) = 
#>   2.3209  7.7964  2.7059  1.9696  6.4705
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]