Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -10.6034   8.6978 -10.5370   4.1647  -4.5180
#>   -1.5150   1.9627  -4.2523  -3.9852   0.8784
#>    8.6365 -12.5353   9.3773  -9.0738   2.0328
#>  -10.4887   3.7392  -4.6713  -2.0609  -5.0839
#>   -0.9513   1.2805   2.7783  -4.3602  -7.9794
#> 
#> (1,2,.,.) = 
#>    1.3037  -5.1838  -2.3435  -9.6398   0.2616
#>   -8.0835  -1.3035   0.9562  -5.9230  -1.1162
#>  -13.4091  11.1629   9.3894   1.5760   4.8357
#>   12.0444   0.1737  -0.6722  -1.2149  -0.9290
#>   -7.2035  -7.9380   2.2488   1.9287   7.4121
#> 
#> (1,3,.,.) = 
#>   8.5760  6.2465  1.3475  3.1817 -3.6854
#>  -9.1300  2.5883  1.2816 -1.3816 -0.1648
#>  -6.2570 -9.2616  1.7773 -4.1021 -0.6669
#>   3.7328  7.6658 -8.0977 -7.7770 -1.7483
#>   1.3393 -3.5501 -0.2025 -0.2943 -2.0280
#> 
#> (1,4,.,.) = 
#>   2.5764 -1.2224 -1.2925 -4.7465  0.4112
#>   4.3375 -2.5581 -7.2546  0.9567  8.3104
#>  -3.5811  1.2463 -10.5706 -4.5739  3.3698
#>   3.4297  5.8039  0.8931 -2.1603 -2.5963
#>   4.9345  0.6983  2.0555 -0.7157 -0.4611
#> 
#> (1,5,.,.) = 
#>    6.0319  -2.2512   2.1890  -5.8183   8.5505
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]