Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -0.7685 -3.8851  5.8796 -0.0979 -7.7383
#>  -8.6995  4.3678 -5.9755 -1.0913  1.7957
#>   2.3370  5.9637  4.8418  8.9690  1.1880
#>  -5.7389  6.8401 -1.1698 -5.0261  6.1673
#>   3.9761 -2.5353  1.6893  7.0093 -1.1964
#> 
#> (1,2,.,.) = 
#>  -0.8747 -0.4549  6.1076  6.8144  0.3218
#>  -2.4097  4.6688  1.6486 -4.6360  8.6163
#>  -0.6197 -6.9600 -3.7749  1.5318 -2.8005
#>  -2.4125  3.0409 -5.4757  2.3733  1.1206
#>   0.7340 -3.7153  3.0781  2.0610 -3.3367
#> 
#> (1,3,.,.) = 
#>   -1.4263  -3.4125   2.7488   2.5853   0.4957
#>   -4.9661  -5.8004  -0.7553  -4.6006  -2.5674
#>    0.5605  -1.7156   2.3695   1.1221  -2.1552
#>   -0.9920   8.9319  11.5179   1.0034   4.8710
#>    3.7614  -0.9369  -2.6405   0.1258   6.7716
#> 
#> (1,4,.,.) = 
#>    1.3432   0.7022   3.0529   1.6414   3.1429
#>   -3.8557  -2.0572   4.1595   2.4989  -1.5509
#>   -4.6023   2.1239  -1.8011   2.4756  -0.8063
#>   -4.8272  11.0755   0.0515   5.0331   1.8820
#>    2.4348  -3.4011   8.6551   2.6122  -1.4155
#> 
#> (1,5,.,.) = 
#>    3.0972   0.0318   5.0399   5.4345   0.4764
#>   -2.0315   5.6252  10.6420  -4.0869  -0.7289
#>   -0.1890   4.8337  -2.1604  -0.0226  -7.8249
#>   -0.8635   5.9217   0.8106   1.7936   9.8291
#>    1.5468  -4.9016  -8.6049  -7.4717   4.2027
#> 
#> (1,6,.,.) = 
#>  -2.3712 -0.0958  0.4412  1.4674  5.4646
#>  -5.3652  3.1997  0.0674  3.1802  1.3037
#>   7.8403 -2.7675  5.5444 -2.1446 -5.9275
#>   1.8768  3.9490 -0.9301  4.5838  2.2543
#>   0.1873 -2.0336  3.6854  2.8355  2.1853
#> 
#> (1,7,.,.) = 
#>   5.5793  2.6985  8.0336  1.8022 -8.3717
#>   0.9055  9.2263 -3.5966 -5.2057  2.0497
#>   6.0702 -1.3387 -1.4833 -9.6995 -5.1903
#>   3.7135 -0.7768 -3.8439 -3.3579 -6.0132
#>   1.9366 -4.3542 -6.3925 -0.9480 -6.6969
#> 
#> (1,8,.,.) = 
#>   -0.5686  -2.1076   3.9401  -2.5111  -2.6371
#>    4.5795  14.8971   2.4927   4.5828   7.0306
#>    0.3333   2.7437  -3.5236  -5.0075  -3.3823
#>    2.7135  -3.7277  -2.0668  -2.9534   3.8298
#>   -2.3920  -7.9266  -0.2248   0.8075  -6.5505
#> [ CPUFloatType{1,8,5,5} ]