Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   5.1498  9.0488  0.7114 -3.4477 -4.3690
#>   3.1287 -2.2929  1.0670 -0.7263  0.5110
#>  -3.5667 -6.9239 -2.9618 -2.5160  3.3889
#>  -4.1963 -2.4198  7.0445  6.5960 -3.5691
#>  -0.4278  2.9538 -9.0968 -1.4667  0.7781
#> 
#> (1,2,.,.) = 
#>    4.5291  -0.6835   2.4213  -5.9263  -3.1530
#>   -5.3579   3.2003  -2.8568   6.8941  -1.8757
#>    2.0841  -0.1575   0.4808   1.1709   1.0182
#>   -7.9307  -7.3222   6.5178 -11.2148   0.8606
#>   -1.1125   1.2498  -4.3776  10.0683   2.5059
#> 
#> (1,3,.,.) = 
#>  -4.3529  1.2870  2.4637  4.8455 -1.3252
#>   5.9436 -11.0837  2.7046 -1.9665 -5.9793
#>   0.6426  7.4859 -0.3309  6.2676 -3.8714
#>  -7.9774  3.7210 -2.2759  4.0371 -2.1748
#>  -4.2186 -1.5034  4.2999 -7.0858 -3.4932
#> 
#> (1,4,.,.) = 
#>  -1.9569 -0.5194 -5.9464 -3.9011 -3.2053
#>   1.3499  9.7986 -3.4548  3.2393 -4.8234
#>  -2.2931 -7.7535  0.8882  1.4256  4.5515
#>  -8.0304  6.3166 -7.1300 -9.1171  0.4058
#>   0.1244 -1.1728  7.4933  0.0983  3.3248
#> 
#> (1,5,.,.) = 
#>    2.2775  -2.6860  -7.0921  -1.7957   2.8886
#>   -1.9353  -2.7752   9.8873   6.6213  -0.2406
#>   -1.3453  -3.8215  -2.7734   0.0394 -11.7802
#>    2.3268  11.5253  -0.2761   1.3072   7.1441
#>    0.7201  -5.6574   5.8988  -5.7547  -4.2908
#> 
#> (1,6,.,.) = 
#>  -2.1384  7.5882 -0.0215 -0.7212  3.6452
#>   0.8753 -10.5653  1.7619 -5.6962  2.6222
#>   3.8563  8.6885  4.1504 -3.1391 -5.7654
#>  -0.5874 -5.2199 -7.7469  0.6477 -3.6576
#>   1.0410  9.4497  5.0282  3.3368 -4.3816
#> 
#> (1,7,.,.) = 
#>    2.1839  -0.7905   2.1986  -1.4281   7.8311
#>   -1.4574  -1.2165  15.0503   4.7291   4.5359
#>    0.6941  10.6012   6.9289  -6.5722  -0.8022
#>    2.4580  -0.2080   7.2409   0.4077   5.0115
#>   -1.1937  -4.2930   6.1438   1.1579   1.0478
#> 
#> (1,8,.,.) = 
#>    8.7141  11.6863  -0.1954  -5.7787 -12.5744
#>  -14.0300  -6.9354 -11.0359   0.2871  -1.1305
#>    4.9302  -6.2475   1.3486  -1.7982  -3.9310
#>   -4.3425   2.3345  -0.9887  -4.9316  -3.4690
#>   -5.3330   0.5670   1.4352   1.6049   4.9472
#> [ CPUFloatType{1,8,5,5} ]