Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    0.4614  -0.4435   8.2211  -4.0894  -1.3083
#>   -2.5235   3.5839  -4.8801  -9.5436  11.3314
#>    0.7704   8.1952   1.0729  -1.8555   5.6519
#>    5.7317  -5.1684  -1.3246   3.9441  -1.3035
#>    1.6107  -6.1823   1.1490   0.6163  -1.2785
#> 
#> (1,2,.,.) = 
#>  -2.5143  0.2938  8.4563  0.4323  3.6732
#>  -2.4596 -0.7728  4.5647  6.4952 -3.9730
#>   5.1863 -2.2055  1.8885 -0.4442  0.0084
#>  -3.8611  3.6556 -1.8884 -0.7836  0.9812
#>   0.5144 -6.3565 -1.3831  6.7877 -2.4959
#> 
#> (1,3,.,.) = 
#>  -6.2933  5.1853  1.8189 -6.9964 -0.8964
#>  -3.0664 -4.4459 -5.7713 -2.8161 -6.0831
#>   3.8812  5.6574 -4.6437 -6.7095  2.2125
#>  -0.6864 -14.2940  5.5942  1.7531 -6.4052
#>  -5.3061 -0.8595 -1.2491  3.9154  1.0879
#> 
#> (1,4,.,.) = 
#>  -0.6393 -14.4735  6.4531 -2.2029 -1.9598
#>  -5.4203  6.3183 -12.9010 -12.2507  2.7948
#>  -6.8282 -13.1412 -2.6293  0.3465 -7.8926
#>  -0.5119 -8.0114  4.8435 -0.6575  2.3413
#>   7.2888  3.3329  4.8111  0.0249  2.5371
#> 
#> (1,5,.,.) = 
#>  -1.7060 -3.2777  4.1970  0.5707 -5.8369
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]