Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -1.9276  0.3060  1.6092 -10.6577  5.7287
#>  -1.7356  1.4964  0.6688  0.5743  3.7858
#>  -6.1002  2.6894  0.0733 -7.1192  4.3014
#>   2.1611 -2.3062  7.6135  4.6696  1.9027
#>   1.4315  3.7313 -0.9806 -3.9600  3.1731
#> 
#> (1,2,.,.) = 
#>  -1.4939 -7.4352 -3.7675 -7.4212 -5.8515
#>  -0.3201 -3.6437  0.0941 -10.0422 -0.6460
#>  -1.0166  1.3953 -1.3316 -3.0048 -11.7465
#>   0.1833 -0.3744 -2.5920  1.9919  7.0045
#>   8.3559  0.0322  6.6055 -3.4513  2.7838
#> 
#> (1,3,.,.) = 
#>   -0.1190  -2.7878  -5.4482  -5.6486  -4.9040
#>    3.6939  -7.4200   2.9213 -10.3045  -2.7197
#>    7.4923  -1.4985   4.9419  -0.1058   1.9639
#>    7.8347   4.7301   2.7123  -6.2687  -2.1397
#>   10.6885  -4.9847  -4.6556  -1.4191   6.8334
#> 
#> (1,4,.,.) = 
#>   -0.7394  -5.3439  -0.0146  -1.3165  -5.5015
#>    1.5622   6.1556   0.6948  -2.1819  -6.2331
#>    1.4021   2.8468  -3.7600   4.2629  -3.0733
#>   -6.9026  -4.4966   1.4730   3.4706   4.0713
#>   -3.0339   4.6334  12.7507   1.0372   1.0936
#> 
#> (1,5,.,.) = 
#>    6.0325   1.7843   5.2341   5.3846   0.7647
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]