Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   1.9654  1.8824  4.9513 -3.6949 -7.9174
#>   3.9147 -3.7345 -2.0971 -5.3162 -3.7104
#>   0.1947 -17.3086 -2.6512  6.5159  5.7785
#>   8.5056  3.0151 -2.3162 -5.1036 -5.7185
#>  -2.3466 -1.5767 -5.8217  9.4632  3.6169
#> 
#> (1,2,.,.) = 
#>  -1.7037 -0.0322 -7.1909 -3.3130  1.2670
#>  -2.8017 -6.0586  0.7615 -3.1906 -7.2204
#>  -8.4662  9.4720  3.3832 -0.5727 -6.5997
#>  -4.2681  4.3107 -4.8360  1.1557 -2.5706
#>   2.3495  5.1121 -13.2043  2.3101 -1.2529
#> 
#> (1,3,.,.) = 
#>   1.8034  3.0733 -4.2999 -7.6217 -0.9993
#>   1.8943  1.9217  1.6909 -7.5412 -10.1848
#>  -9.6286 -0.8811  8.0980  3.8171  2.1259
#>  -1.1643  1.4467  2.6107  3.4836  0.3167
#>  -2.2035 -2.0025  0.8911 -0.1081  1.7420
#> 
#> (1,4,.,.) = 
#>   -0.1179   6.6902   9.4321   1.4597  -1.1847
#>   -2.2750 -12.9884  -6.7329  10.1675   4.1318
#>    0.6321   4.3324  -1.4092  -4.5427  -1.7512
#>    1.9239  -1.4964  -3.7146  -1.6691   1.0809
#>   -0.9735   1.3127  -3.8639  -5.0245   4.3464
#> 
#> (1,5,.,.) = 
#>   -4.3336  -5.5144   2.1946   2.0205   1.5346
#>    1.2531  18.4618  10.4380  -2.5026  -0.3668
#>   -2.0109  10.9723  -1.8344   4.5436   1.3206
#>    2.2607   7.7057   0.0082   1.5133   0.8098
#>   -2.7424  -2.8632   2.3792   6.6040   5.9230
#> 
#> (1,6,.,.) = 
#>   -2.7727   6.0788   2.0584  -2.0601  -8.2282
#>  -10.0699  -2.6869  -3.5488   4.1342   6.3924
#>   11.2589  -2.6358  -6.5034 -15.5706 -10.9184
#>    1.7953 -13.7883  -2.1983  -1.6432   8.3995
#>   -1.6438  -0.4646   9.1420   0.6472  -1.4649
#> 
#> (1,7,.,.) = 
#>  -5.2661  0.9153  4.8198  4.3790  1.9859
#>   4.3413  1.5029  0.0496  0.1702  3.3341
#>  -2.4080  4.0877  0.1630 -0.1252 -0.3703
#>  -4.5085  7.3320  2.1533 -4.6146 -5.0527
#>  -2.0539  1.2120  1.5606  1.9656  1.6240
#> 
#> (1,8,.,.) = 
#>  -1.6258 -2.6729 -0.0141 -3.6229 -1.3770
#>   6.5711  2.5930  3.1846 -1.7109 -4.8098
#>  -4.2076  2.1706  6.2159 -4.6796 -2.6387
#>  -4.0180  2.4670  1.6930  8.4541 -1.0519
#>   0.2268  5.8923 -4.1057 -4.3696  1.0888
#> [ CPUFloatType{1,8,5,5} ]