Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>  -2.3840  3.5898  8.0539 -3.2226  3.3747
#>   4.0504 -0.7763 -8.7529 -14.0512  1.0263
#>   1.3053  2.1069 -0.1611  1.9097  3.2315
#>   0.1164 -4.9120  4.0074  3.0929  4.2188
#>  -9.4351 -2.2797  0.2705 -1.2392  3.0042
#> 
#> (1,2,.,.) = 
#>    3.0804   4.1315  -4.4260  -0.9294  -1.9091
#>   -7.9720   2.0282 -10.1014  11.9844  -2.1982
#>    4.5074  -8.4747   4.0234  -0.6217   7.9692
#>   -5.4574  -8.1498  -5.7537  -2.6553   2.1807
#>   -5.7714   2.6304   6.5843   2.8442  -2.3576
#> 
#> (1,3,.,.) = 
#>  -4.2045  7.3308  2.6671 -7.1059  1.0608
#>  -3.9324 -6.4958  0.4924 -8.1789  3.8465
#>   4.6471  2.6973  7.5679 -8.8845  7.0805
#>  -5.0542 -7.2175  2.2600 -0.4152  4.5116
#>  -2.0687  4.8791  2.3183 -1.3684  0.6036
#> 
#> (1,4,.,.) = 
#>   -1.8546   0.0070  11.3735  -1.6268  -3.6379
#>    2.8914  -2.1189   6.6895   2.7671   1.4802
#>   -3.4332  -0.8565  -2.2671   3.4308   3.3604
#>    5.9770  -3.5930  -3.9469  -0.5626   0.5334
#>    2.6210   4.3178   0.4329   1.3332  -0.9536
#> 
#> (1,5,.,.) = 
#>   2.7339 -1.2337 -2.1692  4.8159 -0.2580
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]