Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   0.9245  1.8494  1.7871 -3.1130  5.0446
#>  -2.6524  7.0200  2.6398 -9.7720  2.8416
#>  -4.1209 -0.9829 -6.4330 -0.9147  1.4706
#>  -8.8678  9.8212  2.3544  8.7245 -5.6684
#>  -8.4434  2.2255 -5.8575  4.9711  3.9497
#> 
#> (1,2,.,.) = 
#>   0.9820 -7.9889  0.3013  0.6695  6.8985
#>   3.6740 -2.7055  2.3513  3.6906  5.3382
#>   9.0659  2.5256  0.3707 -3.5264  6.1008
#>   7.7719 -5.8318  3.0601  7.4352  2.5146
#>   8.3618  0.7467  4.0693 -1.0509  1.3316
#> 
#> (1,3,.,.) = 
#>   0.7912 -6.4437  5.6083  0.2798 -3.5575
#>   1.7871  3.6909  2.0725 -12.8330 -7.1985
#>  -1.2233  3.2086 -1.6209 -6.9685  1.8132
#>  -0.2734  4.9461 -1.9581 -1.1078  9.9447
#>  -0.4299  2.2653 -3.0269  5.6629  5.8259
#> 
#> (1,4,.,.) = 
#>  -0.3250 -4.9569 -4.2657 -6.4873 -2.4452
#>   1.9433 -5.5946  5.1562 -11.8677  4.3393
#>   1.5414 -0.4967 -0.8961 -7.2631  6.7092
#>  -8.1891  0.6778 -0.6898 -6.6547  0.4561
#>  -2.4446 -0.8229 -2.5225 -9.3267 -2.0208
#> 
#> (1,5,.,.) = 
#>   2.6230 -8.1514 -2.5746  4.5198  0.5695
#>   2.1925  6.9327  1.6421  1.1090 -1.8551
#>   1.1297  2.3129  2.6084 -2.6335  3.9247
#>  -0.2011  2.5465 -1.4860 -5.4455 -6.3566
#>   0.7786  4.8756 -2.1026 -1.4971  4.2136
#> 
#> (1,6,.,.) = 
#>   0.2540 -7.9705  6.3497 -7.7109 -2.9602
#>   1.1422 -1.0166  2.1924  1.6941  2.6640
#>  -2.9373 -6.8055 -5.7431 -2.4343 -1.8506
#>  -2.5505  1.4353 -6.2238 -6.9476 -0.2649
#>   3.4749  2.6668 -6.3751  1.0982  5.0387
#> 
#> (1,7,.,.) = 
#>   2.5046 -6.8888  4.7414  4.5292 -7.9995
#>  -1.2129 -1.9904 -1.5621 -2.7074  3.1499
#>   4.8067 -1.7659 -5.3884  5.8314  0.9904
#>  -6.5156 -1.1777  1.8494 -2.4265 -0.0666
#>   2.9109 -2.7087 -5.1846 -7.0616 -3.4466
#> 
#> (1,8,.,.) = 
#>   2.7730  2.8068  5.8258  2.7300  6.3652
#>   4.2693 -7.2291  7.6661 -8.2518 -3.7651
#>  -1.2110  1.7363  8.7316 -10.1096 -2.0732
#>   4.1875  5.5233  7.0323 -2.7296  2.1053
#>  -7.0570 -4.0741 -2.2526 -5.4607  0.8121
#> [ CPUFloatType{1,8,5,5} ]