Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   4.3479 -1.3767 -3.4445  3.0277  3.4604
#>   3.8794  2.8598 -0.2736  1.4168  2.5951
#>   2.0492  1.7051  0.0730  9.5490  1.0923
#>  -4.2063 -1.3305 -0.7254 -2.4171 -7.1093
#>  -6.1847  2.5966  6.6330  6.0754 -1.7744
#> 
#> (1,2,.,.) = 
#>  -3.4370 -6.5673 -1.7385 -1.8653  1.6364
#>   7.3598  4.6923 -4.4187 -5.5638 -0.5656
#>  -2.7378  8.9783  4.2556  7.5003  3.4232
#>  -3.2224 -4.0995  2.1852 -4.5003 -6.7049
#>   9.7503  4.7956  3.9749  7.7955  2.3047
#> 
#> (1,3,.,.) = 
#>   -8.6111  -2.6257 -12.0265  -7.1625  -3.9059
#>    2.5931   6.1548  -9.4596   1.0680   0.9836
#>    1.6257  -5.0004  -4.9045  15.1044   1.0761
#>    0.9986 -14.9164  -7.4347  -7.8637  -6.4829
#>   -8.3801   0.5648   4.0861  13.7058   5.5612
#> 
#> (1,4,.,.) = 
#>   -1.0242   1.4943   6.5971  -4.1091  -3.2923
#>    2.8805   3.5227  -3.9697   0.7771   1.3505
#>   -4.5714  -5.6646  -1.5252 -11.0424  -5.8815
#>   -3.3714  -1.2473  10.6286  -1.0338   0.2775
#>   -3.2076   0.8667  -9.5396   7.6251  -1.7221
#> 
#> (1,5,.,.) = 
#>   -1.7054   3.2340  -3.9313  -5.4682   2.9815
#>   -2.3754  -7.4703   2.9113  -1.4997   2.9345
#>   -5.4287   9.0977   8.0070   1.9803  -0.1179
#>    0.1742   2.3178   0.2136   3.2698  -1.4208
#>   -0.1585 -11.2781  11.1762  -5.0951  -0.4647
#> 
#> (1,6,.,.) = 
#>   1.9067 -3.3731 -4.6206 -4.5634 -4.7788
#>  -12.7361  1.5195 -1.2905  0.1653  4.2771
#>   0.1203  0.6103 -0.0767  4.6749 -1.4103
#>   5.7292 -2.4442 -0.8351  0.2696 -1.7559
#>   7.2613 -9.2274  0.0864  0.4614  7.1957
#> 
#> (1,7,.,.) = 
#>    5.5931   4.0758   2.6048  11.3307   0.0796
#>    7.6428   2.5321   0.3231  -6.0967  -0.8946
#>   -1.1240  -3.0628  -8.7822  -1.8861   5.7177
#>   -9.9965   4.9029   0.2755   6.9795  -5.7552
#>    5.5495   0.9303  -5.1039   1.6400   9.6440
#> 
#> (1,8,.,.) = 
#>  -5.2250 -0.5305  1.6965  2.1566 -8.0205
#>   7.1451  0.4233 -4.2996  1.9542  2.8541
#>   8.0322 -3.0481  7.1281  7.1032  4.5920
#>   2.5398  5.5056 -2.0070  3.3260 -7.7701
#>   0.1408 -7.7161  4.4192  2.8162  4.2198
#> [ CPUFloatType{1,8,5,5} ]