Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -1.4588   5.6425   2.5982   3.6771  -3.8413
#>    3.3734  -8.4908  -5.4220  -3.2818   5.5236
#>    5.7958   0.2945  13.3877   6.6206   1.8235
#>    4.0042  10.0718  -5.0933   2.7953   0.4321
#>   -0.6783  11.4547  -7.9693  11.4278   2.4876
#> 
#> (1,2,.,.) = 
#>  -0.5506  4.1221  4.8731 -2.2661 -5.6249
#>  -0.8343 -7.2453 -13.6659 -5.1650  5.7289
#>  -0.7687  5.2095  7.0076  0.3276  2.2023
#>   0.1583 -5.2686  0.8487 -2.0304 -3.3841
#>  -0.0175 -0.3452 -0.7068 -0.8208  2.6117
#> 
#> (1,3,.,.) = 
#>   -4.7180   7.5342   2.3930   9.4523   3.5226
#>   -9.5577   6.7859   0.7821  -7.2381  -7.1583
#>    1.0947 -10.3417  19.9962   1.6409  -1.7766
#>    4.3987  -8.8221   3.6145  -2.0256  -2.6021
#>   10.7015   1.5898  -0.7120   5.9040  -3.9350
#> 
#> (1,4,.,.) = 
#>   -3.0740   2.7164  -1.5375   0.2387  -3.6228
#>    1.6237  -0.6105   3.3680  -2.7506  -2.6075
#>    9.7171   7.7973   4.5497  -2.7159   3.6746
#>   -2.4873  11.4266  -9.8223  -8.3421   2.6232
#>   -0.4280  -5.0607  -3.7283  -5.3391  -4.1897
#> 
#> (1,5,.,.) = 
#>    0.6104  -2.4699  13.8400   8.1259   4.6573
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]