Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -5.0468 -3.6244 -5.3588  0.3438  5.3431
#>   4.6436  6.0055 -1.9632  3.1865 -3.8969
#>  -10.9799 -5.0304 -0.6709 -1.1093  3.7876
#>   4.7764 -0.2093 -3.1207 -4.0180 -10.2474
#>   0.5094 -8.7951 -7.1867 -10.6873  1.2353
#> 
#> (1,2,.,.) = 
#>   -1.3973  -1.9991  -5.9702   0.2682  -0.0999
#>   -5.3525  -2.9741  -3.9792  -4.4622   3.2723
#>    9.2978  -3.3013   3.4908  -3.8516 -11.8180
#>    7.5354  -1.0520   2.4501  -3.2824   5.9842
#>    5.7248  10.5134  -0.3681   2.1192  -6.0808
#> 
#> (1,3,.,.) = 
#>  -3.7605 -4.1640 -7.1384  2.9435  8.8250
#>  -0.0421  0.6009 -18.6171 -7.5711 -3.9016
#>  -6.1324  2.0892 -3.1568  3.7135  1.9793
#>   2.9089 -4.8238 -1.7748 -0.1700 -5.7891
#>   6.6647 -7.4790 -2.7139 -9.4946  0.9932
#> 
#> (1,4,.,.) = 
#>    1.8850  -3.4575   5.7018   0.3012   1.0212
#>    0.0235  -6.9516   0.1544  12.3822  -0.6555
#>   -2.1066  -9.5180  -1.9588  -3.2501  -5.3850
#>   -6.1750   9.0666  -8.7899   3.9861   1.4311
#>   -0.5270  -1.1184  10.5647  -0.1678  -1.7741
#> 
#> (1,5,.,.) = 
#>   2.0129 -4.0428 -6.9092  7.2487  2.3231
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]