Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -1.3702   0.2529  -1.3237  -0.9858   4.7225
#>   11.2289   5.0502  -5.7150   3.1710   5.0956
#>   -4.6056   6.2880  -3.5991   4.4295  10.5120
#>   -1.8173   0.0950   4.3746  -1.9132  -0.0949
#>    4.8793  -0.0227  -2.0554   0.4306  -1.5657
#> 
#> (1,2,.,.) = 
#>   0.9174  4.2077  0.6846 -2.8909  2.4093
#>  -9.0041 -1.0918  0.9956 -1.2294 -1.5413
#>   2.1040 -15.3265 -8.3456 -8.0021  3.7983
#>  -0.4832  1.5486 -7.9951  3.6906 -0.3225
#>  -2.8445  9.3729 -6.9326  1.8615 -1.6887
#> 
#> (1,3,.,.) = 
#>   -1.8029 -10.8549   4.3694  -1.3743   0.6703
#>    3.7189   1.6270  -0.2022  -4.3822  -7.2630
#>   -4.1254  -3.2476  -3.2989  -3.8596   4.0863
#>   10.0778   4.8933   7.7520  -8.2453   0.9139
#>    0.4048   0.2576  -4.5396   8.3001  -2.9756
#> 
#> (1,4,.,.) = 
#>    7.3875  -8.0247   0.6354   6.0703  -2.1994
#>    4.7996   2.7069  -3.5719   2.1177   0.9158
#>   -1.2830  11.0552  12.8157   0.1396  -0.8035
#>    1.2022  -5.0086   5.8295   6.4249  -3.4195
#>    5.4046   7.2223   2.7348   2.7660  -2.7034
#> 
#> (1,5,.,.) = 
#>   5.9372  5.4132  2.2739 -1.2996  0.5335
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]