Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   0.0562  3.0379 -1.4646  3.1714  4.3795
#>  -1.4666 -3.8472 -0.3646  3.4538  2.4135
#>  -6.1882 -5.7958  2.9362  4.1484 -1.4558
#>   9.2715  6.6907 -4.1777 -4.6983 -3.4703
#>   2.2100 -0.5631  5.5708 -3.8494 -0.6886
#> 
#> (1,2,.,.) = 
#>  -0.6539  8.7254 -2.9305 -6.5480 -1.0717
#>   1.2421  2.4117 -9.5816 -13.0534  1.0514
#>   6.3812  1.5671  3.8290 -4.2061  5.5050
#>   1.0910  1.3794  6.2021  5.9523  3.9357
#>  -6.4624 -3.0963 -0.4851  1.5876 -0.7192
#> 
#> (1,3,.,.) = 
#>   8.5520  1.1455 -1.8526 -4.2435  0.2006
#>   7.0105 -2.9662 -9.2701  0.1011 -0.6270
#>  -3.8611  0.3845 -9.0332  1.5599  0.9073
#>  -2.7112  7.3377 -5.3753 -0.2259  2.7014
#>  -1.1118  2.6669  0.7262 -3.3585 -0.3803
#> 
#> (1,4,.,.) = 
#>   -7.6375  -5.4387  -6.3006  -1.2301   4.5673
#>   -9.5034  -2.9744  11.6019   3.9723   7.6590
#>   -3.7738  -4.6226 -12.6143  -8.1014  -6.8205
#>   -1.3535  11.5255  -3.2336   2.7900  -6.4468
#>    0.0909   7.1312  -2.8073  -2.8751   1.7969
#> 
#> (1,5,.,.) = 
#>   -1.1815  -3.3537  -2.8506  -1.5277   2.0346
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]