Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    0.8108   1.7880  -6.6538   2.8917   1.6200
#>    4.2925   4.8257  -5.8230   0.8311   2.4264
#>    7.1369  10.7506   6.4604  -2.8550   2.7331
#>   -1.6820  12.5874   3.6452  -5.6245  -1.1588
#>   -5.6389   2.8288  -2.1666   1.9184   0.5559
#> 
#> (1,2,.,.) = 
#>    0.7957   1.2093  -4.1477  -3.8764  -2.0031
#>    2.6289  -4.5104   7.1864  -3.2459  -2.5799
#>   -4.6350  10.4554  -2.6483   5.7132   4.4545
#>   -9.0351 -16.0795   0.8554   2.3345  -4.1375
#>    2.8965 -12.8290  -3.3348  -2.6246   0.7263
#> 
#> (1,3,.,.) = 
#>   2.7218  0.4712  8.0682  7.7507 -1.6676
#>  -2.2751 -11.4871 -2.5265 -3.3053 -1.5634
#>  -0.3010  5.2293 -5.4837 -10.3266  4.5894
#>   2.9851 -5.8970  4.4606 -1.7777  1.9251
#>  -2.9132  6.7404 -0.4935  0.7457 -2.1312
#> 
#> (1,4,.,.) = 
#>    4.2211  -3.6396  -2.2822  -3.4153  -0.7976
#>    1.2232   9.6101   4.9518   6.5435   1.7012
#>    7.8707   0.2192  -7.1050   3.1758   1.6313
#>    9.8960   5.5901  10.2468  -6.8787  -5.0413
#>    6.8129   5.9073  -0.5358  -3.3405   2.0544
#> 
#> (1,5,.,.) = 
#>   3.3904 -2.5936 -2.3447  4.8206 -0.1075
#>   3.4734  3.5575  3.3311 -1.9681 -0.0539
#>   5.2319 -4.9462 -8.3794  6.0630 -3.5077
#>  -3.1619 -1.7268  6.5854 -8.7446  3.4354
#>  -0.9644 -4.5135 -1.1772 -2.0911  4.1900
#> 
#> (1,6,.,.) = 
#>   0.6959  0.6227  3.4859  1.3365 -0.3688
#>  -12.0035  5.1887 -2.3511 -0.8687  3.3485
#>   6.9443 -5.9008  0.9329  4.7339  3.3460
#>  -2.6998  4.5250  1.0897 -2.4147  0.9846
#>  -9.4414  7.5349  2.5179 -3.4316  1.4765
#> 
#> (1,7,.,.) = 
#>  -0.3312  1.8130  1.7456  7.5834 -0.2508
#>  -7.1260 -4.0095 -3.3899 -7.7880 -1.0377
#>   2.9493 -7.9712 -2.0885  7.7827  0.6519
#>  -2.7323  8.8935 -10.7014  6.5146  6.2700
#>  -1.1659 -8.1057  0.8686  5.7503 -8.0691
#> 
#> (1,8,.,.) = 
#>    5.3405   8.1894  -9.1693  -0.8018   1.8112
#>   -0.3494   3.8207   2.7967   5.6203  -0.3133
#>    2.2770   1.7621   7.8896   3.2639  -0.0170
#>   -3.1973  12.3725  -6.3146   2.7151  -0.4229
#>   -0.4725  -0.6089   2.9907  -6.4909  -0.8397
#> [ CPUFloatType{1,8,5,5} ]