Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -12.3591 -11.9633   2.5646   2.1545   2.1118
#>   -2.9909   1.6506   2.3760  -1.3058  -6.0348
#>   -1.4479  -3.4556  -3.0999  -0.0415  14.0806
#>    5.6541 -11.7248   4.1567   0.0414   0.2192
#>   -7.7519  -1.3039   5.5531   2.3179  -3.6415
#> 
#> (1,2,.,.) = 
#>    1.7067  -0.3198  -6.4483  -5.6632  -6.6967
#>   -0.3796  22.5576   8.2737   3.5873   5.0727
#>    2.9702  -5.5439  -2.2201   5.9695  -3.1478
#>    1.5374  14.7586  -0.7705  -2.6822  -9.4167
#>   -3.1251  -0.7903  -1.1125   5.1988   4.7339
#> 
#> (1,3,.,.) = 
#>    4.8503  -2.8364  -0.4521  -2.3108   1.8585
#>    2.6196  19.8682   0.9426  -2.4848   2.0962
#>   -0.3904   0.4117   7.9511  -0.6972   1.0749
#>   -0.2745   6.6232  -5.7723   3.3877  -3.7938
#>   -0.9099  -0.1104  -0.9604   3.1565  -5.5898
#> 
#> (1,4,.,.) = 
#>    1.9455  -2.9592   1.5204   0.9392  -4.7037
#>   -4.4472   2.6496   4.1409  -2.5952  -3.3860
#>   -6.4812  -4.3884  13.5774   1.7719   2.4184
#>   -1.1776   8.7583   5.8249   5.1001  -1.1382
#>    9.8450   8.6584  -9.5544   2.0985   3.4280
#> 
#> (1,5,.,.) = 
#>   -3.0579   0.3343   2.7389  -5.4149  -8.2914
#>   -3.5283  -6.5028   3.1036   1.9904  -1.0137
#>    4.1224  -5.7890   0.5421   7.2807  -4.7632
#>   11.7553   9.9592  -3.9218  -1.5322  -0.4629
#>    5.1884   8.8892  -1.4885  -6.8156   5.0262
#> 
#> (1,6,.,.) = 
#>   3.5925 -0.0284  2.9427  7.2698 -0.5787
#>  -1.5843  0.9561  7.5027  5.0638  4.8215
#>  -0.1789  0.3259 -0.2833 -0.6419 -3.9490
#>  -1.2162  1.1611 -8.4390 -0.6506  0.5698
#>  -3.5128  2.5694 -1.7301  1.1001  2.3584
#> 
#> (1,7,.,.) = 
#>    3.5340  19.3890   0.5626  -8.3508   0.7576
#>   10.4750   6.7241 -11.9783   9.7082   1.1820
#>  -10.8751   7.6888   9.6149   0.8048  -6.6914
#>   13.0167 -12.7010  -5.3755  -2.8414   2.2672
#>   -8.4986   9.3796   7.7703  -6.6692   2.0058
#> 
#> (1,8,.,.) = 
#>   -1.6000   1.0876  -6.1010  -8.5210  -5.9104
#>   -0.6735   8.6483  -8.0170   2.7946   0.6683
#>   -6.3325  -1.5930   5.2819  -9.0073  13.7671
#>    3.1630   0.1750   9.9859  -6.3644  -4.9629
#>    2.1767  -2.6456  -0.9638   3.5754  -1.0795
#> [ CPUFloatType{1,8,5,5} ]