Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   8.0881 -2.3066  8.9953 -4.8752 -1.3738
#>   0.5132  3.1154  6.7581  3.9493 -3.4724
#>  -0.6130  2.2542 -0.3656 -0.5723  2.3718
#>   7.3968 -1.1690  8.2829  9.5206  6.3700
#>  -5.3526 -2.9244 -1.5860 -4.4549  9.1243
#> 
#> (1,2,.,.) = 
#>  -1.1316  0.2569  5.1694 -6.8088 -6.0124
#>  -3.4262 -0.6676 -2.1070  0.0231 -4.1402
#>   2.8490  0.0575  0.4139  8.2784 -8.8099
#>   1.6405 -8.6887 -5.5393 -1.6306  4.2321
#>  -1.1652  3.0278 -2.0604 -1.3239  0.2805
#> 
#> (1,3,.,.) = 
#>  -4.6322  4.1756  9.6061 -5.5354  5.7587
#>   0.1890 -5.3629  1.4428  0.5688 -6.7764
#>   3.1456 -0.2533 -1.5696  4.6384 -2.0002
#>  -2.0628 -4.3664  0.9849 -4.7021  3.9201
#>   4.4891  2.3742 -7.6579  1.6656  2.0424
#> 
#> (1,4,.,.) = 
#>  -2.0089 -0.8491 -2.8719  1.2572 -0.2609
#>  -0.8943 -4.0635  4.1631 -3.1653 -1.4982
#>   1.5904  5.3162 -1.1125  7.5011  3.4959
#>  -1.8836  0.4653  2.5525 -5.0843 -3.5149
#>   1.8371 -5.7583  1.3891  3.3619  2.6243
#> 
#> (1,5,.,.) = 
#>  -1.8889 -0.9294  3.1777  7.0203  7.4639
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]