Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   1.1249 -5.4238  1.5212 -5.2189 -0.0852
#>   7.1784  3.8772 -1.6607  3.3344  2.4707
#>  -0.7064 -2.4301  5.6421 -7.0249  1.5856
#>  -3.7301  0.9245  0.5501 -0.8592 -3.6417
#>   5.2778  4.8817  2.4240 -4.8302 -3.5358
#> 
#> (1,2,.,.) = 
#>   1.9641 -5.8935 -7.2085  1.3917 -5.5549
#>  -3.7475 -4.2134  2.2706 -7.9754 -0.0296
#>  -9.7853 -4.2670  8.3013  2.8675 -0.5070
#>   8.5032  0.0243 -4.9254  1.2525  1.1543
#>   1.5629  6.5441 -1.7337  6.4887 -4.4514
#> 
#> (1,3,.,.) = 
#>  -5.0708 -2.9246 -2.8683 -6.4098 -3.7123
#>   3.2070 -1.1813  1.4760 -0.3887 -1.7781
#>   4.5436 -2.4499  3.9078 -15.7377  4.9553
#>  -1.7591 -1.7300  3.9322 -3.5260  4.2380
#>   6.7493 -6.9938 -2.9803  4.7764  6.8917
#> 
#> (1,4,.,.) = 
#>   0.9738  3.1660  1.1834  6.5949 -0.6770
#>   2.1342 -2.8137 -6.4160  0.6777  0.5905
#>   3.3886  5.1637  0.5788  7.2250 -1.7986
#>  -1.0328  3.7161  3.9015  4.9383  2.1341
#>  -3.9499  1.8535 -2.4730  1.9184 -11.4698
#> 
#> (1,5,.,.) = 
#>  -0.0240  1.0973 -0.1196 -2.2065  0.0879
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]