Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    1.5009  -4.3473  12.0772   8.9139   2.4320
#>   -1.0328  -0.6237 -15.7218  -8.8957   4.3947
#>    4.1864   1.3342   1.1351  11.7888   5.8438
#>    1.3651  -4.9756  -2.1739 -11.8082  -1.8000
#>  -10.3026  -8.2009   6.1494  -2.2194  -0.9984
#> 
#> (1,2,.,.) = 
#>   -3.3800  -2.1301   0.6847  -2.1040  -0.1006
#>    1.4388   0.1344   7.8547   2.0821   5.5422
#>   -1.7401  -6.6261  -7.3062  10.5552   6.2941
#>    4.5236   4.0829  -8.2950  -0.1537  -1.8011
#>   -3.0351  -0.2508   3.0911  11.9471   1.6782
#> 
#> (1,3,.,.) = 
#>   -4.8648  -0.1853  -1.3084  10.5818   4.9067
#>    8.8654  -7.0407   4.9519  -8.0880   0.5341
#>   -3.4525   0.3839   1.6753  -1.3405   8.3962
#>    1.5189  -0.8353   1.5052  -6.3625  -0.4302
#>   -4.5448  -3.2169   4.7424   3.9297   6.5250
#> 
#> (1,4,.,.) = 
#>   -0.6595   1.6199  -8.5334  12.5902  -4.5060
#>   -0.0622   2.8055   3.1365   0.4630 -10.4077
#>   -4.2585   2.2506   1.2154   0.7123   0.0279
#>   -7.6291   4.6839  -2.3267  -0.5536   0.9046
#>    1.7406   1.3266   1.8529   8.4191   4.1264
#> 
#> (1,5,.,.) = 
#>    5.0687  -5.6558   6.8173  -5.8299   8.5321
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]