Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    4.9623  -1.8122  -0.2179   1.2348  -6.6050
#>   10.9431   2.9435  -0.8830  -2.5020   4.1311
#>   -3.3727   1.5954   3.3437   4.8221  -8.4589
#>    1.4863   2.3396   2.4684  -3.6273   1.9901
#>    6.7618   3.1677   0.8829   1.2149  -5.8137
#> 
#> (1,2,.,.) = 
#>  -7.1137  4.6409 -1.5417 -7.3785  3.3189
#>  -3.6005  4.1105  5.3410 -9.3413  5.4950
#>   8.1577 -1.3851  2.5607  5.8282  5.3141
#>  -8.4706 -2.1193 -0.9171 -9.1645  1.1111
#>   2.8342  8.7862  5.7702 -2.9794 -2.0052
#> 
#> (1,3,.,.) = 
#>  -1.1604 -1.0987 -0.5324  1.5445  2.9996
#>  -7.4902  1.8724  2.3486 -1.4882  2.8294
#>  -2.6420  1.2435  4.3556 -0.7736 -5.5603
#>   7.6693 -1.4605  2.5394 -6.5077  7.9816
#>  -5.3188  0.4096 -3.4648  2.4333 -1.9367
#> 
#> (1,4,.,.) = 
#>  -12.7107  -0.2842   7.5592   6.9539  -0.4070
#>    4.8932  -1.3681   5.9847   0.8260   0.6641
#>   12.0276   2.7923   0.1802   0.2692   0.7227
#>   -3.1187  -0.8582  -7.9721  12.4272  -0.1677
#>    5.4586   1.4613   0.3279  -2.7892   2.6345
#> 
#> (1,5,.,.) = 
#>  -4.1718  0.5010  3.5356  5.3582 -4.4277
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]