Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -1.3075   0.9281   1.3642  -3.1526   7.0591
#>   -6.4464  -4.6577   0.3641   1.1960   2.3680
#>  -11.5164   0.6915   4.9803  -6.9258   4.2176
#>    5.0367   1.3483   4.1329  10.3842   9.3329
#>    1.6954  -8.7699   3.9803  -2.5008  -8.5366
#> 
#> (1,2,.,.) = 
#>   -0.8066   1.1830  -1.4016  -3.2033  -1.4286
#>    1.5912  -6.6994  -4.1122   4.7439   0.4849
#>   -5.6748  10.9473  -0.5484  -8.1948  -5.6207
#>  -11.3308   2.3907  -2.7664  -9.1879   0.1618
#>   -8.7250  -2.1746  -1.3358   0.5535   2.0157
#> 
#> (1,3,.,.) = 
#>  -2.4539  4.4497 -10.0913  7.0147  3.6899
#>   4.0235 -3.6534  7.0059 -9.0249 -1.2551
#>  -0.9512  4.9655 -3.8814  3.5012  0.2836
#>  -5.5645  2.0353  1.1169 -2.9693  8.5852
#>  -1.6881  0.8538 -2.1916 -2.8001 -2.4560
#> 
#> (1,4,.,.) = 
#>   11.7219  -1.8325  -0.7289   1.4364   0.3490
#>    0.4894   8.8353  -3.7560  -0.1113  -1.8818
#>    3.3132   1.3354  -9.7499 -16.5345  -8.5630
#>   -4.9527  -1.4075  -2.6884   3.1978  -1.6261
#>    2.5961   3.8508   0.9623   0.6890  -2.4992
#> 
#> (1,5,.,.) = 
#>   -8.0545   8.9828   6.8685  -1.9843   6.2347
#> ... [the output was truncated (use n=-1 to disable)]
#>   0.2188  0.1843 -0.7550  1.0533  4.5182