Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    4.3374   1.8776   1.7855  -1.6795  -2.5075
#>   -3.5462  -4.5106  -1.2177  -4.6614 -11.4694
#>   -4.8152  -2.2011   6.0428  11.3376   3.8651
#>   -2.4105   3.5251  -2.8189  -0.0835   1.3191
#>   -1.8150   9.5929   1.6609  -4.4090  -3.5676
#> 
#> (1,2,.,.) = 
#>   -5.2561   0.0323  -2.5149   4.8374  -0.0901
#>   -5.8099  -1.8550   3.3631   0.0595  -5.8540
#>    6.9853   1.6576  -0.5564   3.2740   9.2498
#>   -3.0220  15.1662   2.5260  -4.0571   3.4795
#>    0.4223  -1.3955  -1.2241  -0.4612   0.7401
#> 
#> (1,3,.,.) = 
#>   1.6064  5.3332  2.0557  0.9045  6.0357
#>  -3.3206 -2.4351  1.8005 -5.2646 -2.3539
#>  -1.6019  1.7764 -12.0080  1.8922 -1.4449
#>  -4.7969 -1.2249  0.6446 -9.8747 -1.5940
#>  -0.8125  0.0848 -1.6474 -0.1274  1.1765
#> 
#> (1,4,.,.) = 
#>   -4.3450  -4.9715   3.6872  -6.2725  -3.1025
#>    3.7098   0.6209   2.0177  10.0944  -0.6206
#>   -0.9443  -6.9687  -2.9198   7.7043  -9.0046
#>    4.9501  -2.4748  -9.1358  -2.7343   4.0362
#>   -2.4062   1.9036  -2.4564  -1.7319  -4.9160
#> 
#> (1,5,.,.) = 
#>    2.0530  -8.3069   4.8499  -5.7233  -2.3582
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]