Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    0.9017  -3.8809  -2.5431   8.5887  -8.9487
#>   -3.3135  -3.9376   7.1169   0.1856   6.4217
#>   -3.2069  -2.3495   3.4286   0.1707  -2.7285
#>    0.0896  11.0639   1.1676  -6.5264  -0.2672
#>   -0.8227   2.3858  -2.3332  -0.6330   1.6544
#> 
#> (1,2,.,.) = 
#>   3.6328 -2.1009  1.4710 -5.2093  2.2325
#>   2.1386  0.9482  2.0873 -4.9682 -0.2906
#>  -6.0082 -6.7473 -8.9197  2.6142 -3.5233
#>   1.4349  6.2419 -3.7994  6.8772 -3.9377
#>   3.4617  6.8565 -5.6756  2.0390  0.1902
#> 
#> (1,3,.,.) = 
#>    0.0064  -8.6843 -12.4112  -5.6343   2.7864
#>   -2.6036  -7.3524  10.4175  -0.8421   3.3792
#>   -5.5193 -13.5203   4.5406 -10.0838  -7.1015
#>    3.9109   6.2831  -8.3945  -5.6318   7.7389
#>    0.8851  -6.9705  -9.1476   5.7475  -1.2271
#> 
#> (1,4,.,.) = 
#>   2.0432  1.9702 -6.5684 -5.2495  6.1850
#>   0.5362 -7.9832 -2.7527  2.7505 -4.9488
#>  -1.7872 -7.9963 -1.7601  0.4608  1.3226
#>  -0.7097  3.5808 -10.1449  2.9145  7.4469
#>   3.7933  2.4344 -0.7672  2.5211 -1.0293
#> 
#> (1,5,.,.) = 
#>    0.8312  -4.4255   2.6666  -8.9925   4.1471
#>    1.1940   0.5487  -6.7867   0.3734   0.9323
#>    6.7611  -0.5340 -13.8973  11.1292  -1.2449
#>    5.1413   6.1623  -2.7496   2.2521  -1.9586
#>   -4.6021   0.2563  -2.0597   2.1783   0.7686
#> 
#> (1,6,.,.) = 
#>  -14.2010 -10.5502   9.7668 -11.8544   7.6248
#>   -3.2381  -8.7102   5.0010   5.7107  -5.0914
#>   -2.9123   5.0718 -11.6388   6.2908   2.0580
#>   -0.2664   5.6525 -12.0189  -4.4499   1.1538
#>   -4.7629  -2.9188   2.9406  11.0287   0.5413
#> 
#> (1,7,.,.) = 
#>   -0.7921  -0.9243  -6.4910   6.8950  -3.3395
#>    1.2698   5.1437  -2.9131   5.5087  -5.0209
#>   -1.4430  -0.8488  -3.6532   3.0702  -2.7110
#>    2.2078  -7.1707  15.4630 -10.7022  -0.5147
#>    5.6253  -0.0365   2.0614  -5.7122   4.3203
#> 
#> (1,8,.,.) = 
#>   -1.9151   0.1151  -0.7131   0.8088   2.5222
#>    5.1556  10.1108   4.7323  -2.5301   0.1424
#>    6.4479  -8.6761  -1.7588  10.5235  -7.4265
#>   -4.6858  -1.4758   9.5658  -2.0105  -0.3098
#>   -1.8547   3.6239  -4.8406  -7.3639  -2.0709
#> [ CPUFloatType{1,8,5,5} ]