Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -0.1375  1.0286  1.1485  6.5669  5.1903
#>  -3.1714 -6.1459  7.2901  3.5052  1.7800
#>  -3.4140 -1.8644 -3.7778  7.8454 -6.3327
#>  -5.3747 -4.2752  1.3747  2.4004 -1.9506
#>  -3.0685 -3.1213  2.2886 -4.0587 -4.2780
#> 
#> (1,2,.,.) = 
#>  -2.1951 -0.9976  2.5659 -0.2043  0.0468
#>   3.3729 -4.0642 -3.8791 -3.2323 -2.4166
#>   0.9141  1.4438  2.2391  4.1729  0.8801
#>   6.1538  6.0424 -2.7026 -1.5697  1.4883
#>  -1.3065 -2.6134  0.3612 -1.1271 -1.8842
#> 
#> (1,3,.,.) = 
#>    7.7638   5.6758  -0.0575  15.6297   1.5178
#>    6.9608 -10.8565  14.9602  -4.4287  -4.6267
#>    3.3738  -0.2566   1.9011   8.1330  -2.6914
#>   -1.9104   3.9166  -0.8433   2.7178  -4.2451
#>   -3.0903   0.1425   4.7395   0.0413   2.0488
#> 
#> (1,4,.,.) = 
#>  -1.3319 -2.6971  0.7170  1.2580 -8.6731
#>  -2.8630 -10.4183  0.7008  6.4555 -4.0089
#>   0.7445  6.4088 -10.2124 -6.7459  1.4306
#>  -1.7926 -7.7577  0.6959 -4.1776 -5.1696
#>   3.7842 -2.1881 -4.9916 -0.2728  3.0606
#> 
#> (1,5,.,.) = 
#>  -3.7451 -0.1661 -2.8102  4.0113 -0.3067
#>  -1.1514  4.0124  0.2650 -6.6455  5.2233
#>   1.2870  2.6636 -0.6601  2.6084 -0.1338
#>  -1.3285  0.5759  2.6541  7.3344 -8.8096
#>   1.0759 -3.1080 -0.1353 -6.4965 -1.0434
#> 
#> (1,6,.,.) = 
#>   0.0380  2.9989 -1.6935  8.1221 -0.3298
#>  -1.5081  7.5415 -2.8672  2.3662  5.4182
#>  -4.7267  4.7474  3.0339  5.0753 -4.8540
#>   1.6162 -0.1508 -8.0162  5.4801 -6.0751
#>   4.7528  5.0063  1.0713 -8.3608  1.3195
#> 
#> (1,7,.,.) = 
#>   2.4257 -1.5349  2.6383  2.0768 -0.5449
#>   2.9268 -0.7571 -11.6912 -2.2997  1.6379
#>   0.1899  3.6039 -2.5723 -10.2365  4.0532
#>   0.7109 -7.6946 -0.5628  4.4732  2.4435
#>   4.9227  7.1215 -4.3277 -5.7969  0.4140
#> 
#> (1,8,.,.) = 
#>  -3.2949  4.5113 -6.1201 -5.5399 -4.5266
#>  -1.4273  0.3395  2.2981  8.8722 -0.4924
#>  -4.0550 -10.8062 -9.4597  0.4110  0.6113
#>  -2.7646  4.0401 -8.1379 -5.6275 -14.9286
#>  -2.2635 -6.7788 -7.6813  7.5289  4.4735
#> [ CPUFloatType{1,8,5,5} ]