Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -1.3990   4.4925  -3.5125  10.5207  -3.2310
#>    7.2085   3.1341  -1.3261   1.2084  13.1745
#>   -8.2776   4.9146  -1.3940   4.6685  -0.1193
#>    5.5765   4.2892   9.3602  -4.4408  -3.2562
#>    0.6143   5.3702   2.8156  -2.8349  -5.2562
#> 
#> (1,2,.,.) = 
#>   -3.9085  -4.6638   0.6500   1.2276  -2.3761
#>   -4.7852   4.4805  -4.3325   8.3429   1.8449
#>    0.5226  -1.0779  13.8216  -7.8498   7.8492
#>   -2.6911  -7.7209   3.1005  -4.7983   0.4540
#>    6.1423  -7.9606   7.0031   3.9039   5.7418
#> 
#> (1,3,.,.) = 
#>   -7.3101   6.5273  -3.6545   6.3797  -5.0992
#>    0.0126   7.2425  10.9080   9.0724  -5.5911
#>   -2.8130   3.6432 -12.4149  -2.2025  -5.2641
#>   -0.1945   2.5355  -0.2782  -1.6874   4.5685
#>   -1.2468  -0.0215  -0.7785   2.0925   4.8847
#> 
#> (1,4,.,.) = 
#>   -2.6997  11.0674   0.9201   8.8640  -3.0142
#>    3.7024   8.1088  -3.6729  -5.1148 -10.7942
#>    2.4797  -0.2726  -5.9253  -9.8053   0.9098
#>   -9.5870  -6.3336  11.0211  10.3734   7.1479
#>    2.5443   0.7379   7.7906   1.0524   2.4800
#> 
#> (1,5,.,.) = 
#>  -1.7302  2.3235 -0.8379  6.3180  3.8479
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]