Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -0.5359 -4.4082 -0.0579 -1.4095 -3.6971
#>  -0.1008  2.6480 -4.0544 -1.4343 -2.7727
#>  -2.5523  5.0237  0.3433 -2.0039  4.0774
#>   4.9800  3.0548  1.0986  2.8181 -3.1034
#>   0.8432 -1.9117 -3.0844 -0.5877  5.2301
#> 
#> (1,2,.,.) = 
#>   1.2048  4.5210  7.9938  0.3537  2.1726
#>  -5.5794  0.4531 -4.0706 -3.2729  4.5521
#>   8.8866 -7.7700 -3.9268 -3.9051  1.5309
#>   5.1940 -10.4000 -1.3581  4.1340 -9.4870
#>   2.8326 -12.2362 -2.0057  6.1641 -0.6869
#> 
#> (1,3,.,.) = 
#>  -8.5046  0.2766  3.8002 -5.7070 -4.5135
#>   0.8549 -5.7073  3.4253  2.7480 -5.0900
#>   4.3187 -0.1510  0.2585  2.0083  2.0813
#>   3.0473  3.0460 -4.4258  5.8853 -4.2063
#>   8.7506  4.5425 -10.4887  1.9535  2.7639
#> 
#> (1,4,.,.) = 
#>  -2.0385 -3.8343  2.5755 -3.8022 -0.5599
#>   0.3956 -9.3331  4.0774 -3.9226  1.8957
#>   7.8111  2.5521 -7.3918 -3.5918  2.8578
#>  -0.7727 -3.4559  1.8783 -5.3758 -4.4688
#>   5.1981  4.3328 -5.3946  3.7513  5.4180
#> 
#> (1,5,.,.) = 
#>   -4.3546   4.2165   1.2749   1.3305  -2.7980
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]