Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    2.0708  -0.6202   4.2032  -0.7694  -5.4441
#>    3.1753  -1.0967   1.4768  -4.7530  -1.6073
#>   -3.9337 -13.7922  16.0367  -1.8468   6.7990
#>   -4.7662   2.0425  12.7266  -9.9259   0.4940
#>    0.3957   9.8920  13.0916   0.5155   0.0707
#> 
#> (1,2,.,.) = 
#>   1.6201 -1.3396  3.6807 -0.5544  2.2846
#>   1.4647  8.0309  5.9297  1.0622 -0.9496
#>   1.5364  7.1787  4.7170 -11.5182  2.7752
#>  -6.3050  4.4377  0.9502 -5.3094  1.8810
#>  -1.6979  8.3121  6.6601  3.5057  4.0077
#> 
#> (1,3,.,.) = 
#>   1.3307  0.4518 -3.9245 -5.2146 -2.9130
#>  -1.0921 -3.2591 -7.6188 -0.1666 -2.6972
#>  -4.7182 -8.6926  1.9723  1.5008  3.2295
#>  -8.1474 -14.2685 -0.2272 -3.3814 -10.3906
#>  -1.6120 -3.9935  0.2171 -4.0324 -8.5222
#> 
#> (1,4,.,.) = 
#>  -5.4377  4.3118  5.8850  2.6577 -1.2985
#>  -1.6428 -9.0512 -7.9095 -5.6763 -0.6893
#>   4.6802 -2.2667  7.2857 -1.9391 -0.7606
#>  -8.0911 -4.6546  8.1090  4.8473 -5.8827
#>  -2.3769  6.0201 -0.5807 -3.0738  2.9154
#> 
#> (1,5,.,.) = 
#>  -5.1972  3.0156  1.3205 -9.6267 -3.1878
#>   0.3683 -9.5519 -11.5578  1.8934  6.7190
#>  -3.7641 -10.1549  3.4091  2.8382 -8.2295
#>   0.2572 -4.8103  5.8321  2.8306 -0.6151
#>   3.8097 -0.7364 -1.9688 -3.5918 -3.6434
#> 
#> (1,6,.,.) = 
#>  -2.4319 -2.4752  3.4615 -0.8338  0.3474
#>  -5.5675 -1.5654  5.2934 -6.7935 -2.1664
#>   3.0682  7.3297 -6.6991 -2.5439  5.6028
#>  -0.8517  3.3044 -3.5135  8.5589  1.2287
#>  -0.5206  7.1687  4.0333  9.0097  3.7554
#> 
#> (1,7,.,.) = 
#>    1.0215  -7.3594  -7.4159   0.1414  -5.5352
#>    3.1852  -5.3901   9.4264   4.6669   0.1606
#>   -3.7179 -11.3081  -0.8583  -3.1510   2.7909
#>   -1.4430  11.6952  -5.3307  -7.2445   0.0790
#>  -11.0000   5.7108 -12.1046  -6.0666   1.2115
#> 
#> (1,8,.,.) = 
#>    1.5537   0.3706   0.3827   8.4218  -7.1957
#>    5.3778  -3.9552   5.0613  -2.0556  -4.2495
#>    5.2490 -15.4681  10.5618  -0.0399  -2.6742
#>    0.5629  -7.4732  12.3082  -7.3271   2.6336
#>   -1.2862  -2.7514   7.8600  -4.3825  -1.3895
#> [ CPUFloatType{1,8,5,5} ]