Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -0.7963  -3.6647   3.4218  -7.4083   2.9680
#>    1.3169  -4.8378   3.7282  -1.7261   6.7589
#>   -0.2114   8.4390  -6.1311  16.0224  -7.6369
#>   -0.4502   3.2393   4.5103 -14.2243   6.1253
#>    1.9356  -4.6956  -2.6893   4.7728  -3.4443
#> 
#> (1,2,.,.) = 
#>   -3.3629  -2.0351   2.4929   2.1855  -2.7312
#>   -2.2358  -1.4955  -4.3561  12.2338  -1.5566
#>   -1.6365   0.5848   2.8976   7.5763  -4.5127
#>    7.1877   6.8246  10.4777  -4.2743  -2.0520
#>   -3.1537  -3.5927  -2.1710  -7.4418  -2.2956
#> 
#> (1,3,.,.) = 
#>    2.5336   4.8506   1.0158   1.8704   2.0097
#>   -5.3416   4.4427  12.0807   5.5130   6.3614
#>   -0.4009  -0.8074   0.4138  -1.0767   8.8404
#>   -5.5180   1.5930   2.5004   7.5504  -2.1220
#>   -2.2602   5.7687   0.5561  -1.2359  -1.6146
#> 
#> (1,4,.,.) = 
#>  -1.0042  1.4005  3.4752 -0.0227  5.7255
#>   6.7331 -2.9952  2.6170 -4.3030 -2.8938
#>  -7.3975 -1.5889 -8.2143  8.2223 -8.8620
#>  -3.6025  2.1849  6.4135 -2.5471  6.0607
#>   2.9892  3.0244  4.7347 -0.8184 -1.7988
#> 
#> (1,5,.,.) = 
#>  -3.1874  6.3823 -2.5305  6.1862 -4.5020
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]