Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    0.7902  -3.7177   5.5598   5.6339   3.1727
#>   -5.1328  -2.3188  -6.1686  10.8253   7.1079
#>   -0.5177 -10.5782   1.1714   5.9521  -2.5189
#>   -6.1871  -1.2173  -8.2338  -3.7103   4.6158
#>   -4.4132   1.9778  -9.2017  -2.8260   0.6475
#> 
#> (1,2,.,.) = 
#>  -5.3493  3.2604  3.2607 -2.7658  4.0933
#>   6.1684 -9.7086 -1.4645  2.4180  5.7467
#>   2.0567 -9.9704  5.7242  0.4006  2.4180
#>  -1.8001 -5.7582  2.7535 -5.5685 -2.0182
#>  -3.6451  2.8267  4.3085 -0.0756 -4.1572
#> 
#> (1,3,.,.) = 
#>    3.8098  -4.0583   0.6875  -1.8866  -5.8463
#>   -1.2572   6.8911  11.8354   2.4800  -3.1064
#>   10.9991   3.6244   4.0346   7.0628  -0.4687
#>    2.1438  -3.2172  -0.6831  -3.1176   0.3663
#>    5.2558   0.0857  -2.7101  -3.6192  -0.2389
#> 
#> (1,4,.,.) = 
#>   6.4980 -3.6335  1.9141  1.2716  3.8330
#>  -4.5464 -0.3374 -0.7988 -5.3008  3.3030
#>   1.0864 -3.7450  3.6066 -1.7758 -2.8959
#>   3.6041 -11.5954  0.3691  3.9375  2.6004
#>  -5.7053 -0.1314  5.0996  6.7191 -0.5976
#> 
#> (1,5,.,.) = 
#>   -1.1723  -0.6228   1.8764   4.4804  -1.8314
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]