Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   7.2385  3.3794  2.9444 -4.3073  0.4275
#>  -15.3280  9.1316  3.8297  3.3216  4.5066
#>  -0.4931  6.1517 -1.7697 -1.7717 -1.8722
#>  -2.3817 -3.4362 -1.6093 -0.6644 -6.3946
#>   2.2340  4.9381 -6.0945  1.0633 -2.0344
#> 
#> (1,2,.,.) = 
#>   -4.7215  -7.1918   2.0449   3.7430  -0.0706
#>   -5.9504   1.2115 -14.1968  -7.7756  -2.5614
#>    6.6308   1.8676  -4.5335   6.7236  -6.9020
#>   10.6841   3.0293   7.7823   0.5180   3.1116
#>    3.4647  -2.8445   7.0408   1.7338   7.2038
#> 
#> (1,3,.,.) = 
#>   2.4152  7.2217  0.3330  7.7053 -0.0885
#>   5.1478 -6.5374 -2.1261 -7.3168  2.5358
#>  -1.1599 -5.6175  0.2339 -11.7040 -1.8711
#>  -1.1490  0.5736 -4.4246  1.1796 -1.9299
#>   4.9162 -1.8464  0.6853  0.7067 -2.7537
#> 
#> (1,4,.,.) = 
#>  -1.2362  1.2391  2.5635  7.1748  3.9052
#>  -3.2258  2.2962  1.7990  1.5431 -1.6150
#>   1.0314 -1.5055  6.4781  1.4588 -1.7164
#>   7.4284 -3.7815 -1.4993 -6.2661  1.3383
#>   3.7748  4.4881  4.2697 -5.2294  5.2273
#> 
#> (1,5,.,.) = 
#>  -7.9329 -3.7228  7.5973  0.1706 -4.1445
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]