Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -3.9207   4.0716   1.0245  -5.2431   4.8063
#>    8.0709  -7.3844  -5.5230  -1.9289   2.2841
#>   12.2867  -3.4415   2.6815   2.6268   3.9908
#>    1.6049   3.5054 -10.6357  -0.9225   2.0344
#>    1.3502  -3.8906   3.5658   1.2238   3.3591
#> 
#> (1,2,.,.) = 
#>  -0.3302 -2.8395  4.7609  0.8798 -4.4895
#>  -2.2364  6.5816  6.7481  1.1745 -3.7499
#>  -4.0555 -4.0882  7.3022 -6.0248 -4.6221
#>  -3.5374  4.4315 -4.4612  2.5519  1.5247
#>  -0.9968 -2.3962 -2.9624  1.6350  4.3887
#> 
#> (1,3,.,.) = 
#>  -9.0258  3.4909 -3.1228 -2.7185 -1.6285
#>   1.5742 -0.9138  4.2927  3.9369 -5.8129
#>  -0.3960 -7.2321  0.5409  4.7547 -1.5129
#>  -1.2368  9.9749  2.3767  2.4707  2.1366
#>   5.7093 -2.6524  6.1150  0.0055  2.6763
#> 
#> (1,4,.,.) = 
#>   3.6136  2.5665 -1.0160 -5.7810  4.2774
#>   3.5565  0.2432 -4.9639 -12.0859  6.1441
#>   4.5241 -4.5032 -0.9872  8.0083  1.0928
#>   9.8209 -6.0754  1.8196  2.0437  3.6774
#>   1.1787 -3.4991  9.9907  6.3296  4.1842
#> 
#> (1,5,.,.) = 
#>    3.2959  -2.5306   4.3538  10.7111   5.8093
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]