Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   1.4635 -1.3350  0.8038  0.1034 -2.8682
#>  -0.8607 -0.2026  6.6621 -4.5785 -2.5808
#>  -4.6992  6.5198  2.3978 -3.6814 -2.1220
#>  -10.2580  0.8197  8.1501 -5.0954 -2.4528
#>  -1.6305  0.9526 -5.7230 -0.6502  1.8553
#> 
#> (1,2,.,.) = 
#>   0.3216 -5.5154 -0.9571 -1.0302  5.5476
#>  -7.4760 -8.7564  0.5088 -0.2111 -1.6279
#>  -2.2020 -7.7343 -2.2510 -4.8629  1.5267
#>   4.6055 -2.6929  0.3778  5.0689 -5.0959
#>   3.0877  6.2903  2.3184 -7.6690  3.2047
#> 
#> (1,3,.,.) = 
#>   -3.9223  13.5219   9.6594   4.0449  -3.5528
#>    0.7714   0.6804   2.5187   5.0797   1.8162
#>   -2.7894   0.1312   1.5483   7.4338  -3.6008
#>   -0.6493 -13.1250  -2.3238  -0.2456   0.8153
#>    0.3571 -10.8477  -6.8780   4.7612  -7.9102
#> 
#> (1,4,.,.) = 
#>    0.5623   8.0238  -4.5668  -2.4276  -4.1812
#>    0.3408  -1.0017   7.9135  -3.0557  10.3819
#>    0.1489   9.3419   1.4574   2.1017  -1.5289
#>   -2.3389  -3.7731   7.4623  -2.2710   6.4179
#>   -2.7938  -1.4349  -1.1721   2.4045   2.0223
#> 
#> (1,5,.,.) = 
#>    0.3976   3.0974  -1.9656   2.9550  -3.6146
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]