Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -5.4008  -4.0212  -9.7077  16.2300   2.2669
#>    0.2638  -2.4621   8.4159  -2.7899  -2.4408
#>   -1.7304  -6.9934  -2.8095   6.8837   4.3249
#>    2.0771  -3.0957  -2.3973  -7.5209   3.5395
#>   -2.2548   4.4271  -0.3172  -2.2618  -1.4014
#> 
#> (1,2,.,.) = 
#>  -5.0677 -4.4780  1.7028  3.6041 -1.7623
#>   0.9171 -3.6023  0.3599 -0.8592 -0.7773
#>  -6.8352 -0.9308  4.9914 -2.8542  6.6713
#>   2.4455 -0.9329 -6.2243  6.8460 -2.9406
#>   0.9483 -0.6767 -0.6623  5.3575  1.2086
#> 
#> (1,3,.,.) = 
#>   -2.4756   7.4567   7.1446  11.2686   6.6190
#>   -1.3261  -1.1548   1.3917   8.7589   1.6042
#>    0.1851   0.9093   2.5816   1.8328   5.7951
#>    2.3142  -0.9114  -0.3332  -2.5417  -5.7733
#>    4.6561   7.7108   2.2157   5.5791   0.8343
#> 
#> (1,4,.,.) = 
#>   3.0001  0.2778  1.3574 -5.9023 -2.9576
#>  -8.1800 -5.5580  6.7927 -2.8271  5.1259
#>  -0.6847 -3.1746  6.5450 -6.7946 -2.3404
#>  -3.5016 -4.5350 -3.5325  5.3870  4.6863
#>   2.7988 -0.0217 -1.1106 -0.1356  1.0906
#> 
#> (1,5,.,.) = 
#>   -3.3359  -3.3625  -5.3505  -7.9733   4.2010
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]