Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    5.3010   1.2545  11.5211   1.2661   0.3122
#>   -2.6482  -3.3393   2.9592  -0.6888  13.1594
#>    7.2644   2.9543   4.0442   5.2444   9.5314
#>   12.8991   9.5809  -5.9886   0.8809   1.8129
#>   -1.0294   0.3432  -9.9109  -3.0891  -0.0501
#> 
#> (1,2,.,.) = 
#>    1.1700   1.3862   6.8644  -0.6324   0.1871
#>    3.4938  -6.8177  -5.7179  11.2465   8.7941
#>    5.0113   5.6463   8.5168  -0.0524   3.8665
#>    6.6870   2.1243  -0.9644   5.6198  -5.4496
#>   -1.1434   3.2598  -0.9118  -0.0595   3.4676
#> 
#> (1,3,.,.) = 
#>    4.8473   5.9894   4.3534   6.4798   5.2910
#>    5.0279   1.7024  12.0814  10.9086   7.6376
#>    4.1073   0.9622  -1.5559   3.0090   7.1852
#>   10.5846  -1.4466  -8.3569  -1.0206   0.0486
#>   11.5596   3.5966  -0.7361   4.5576   5.6315
#> 
#> (1,4,.,.) = 
#>   8.0570 -9.4673  6.4279  0.5086 -1.9695
#>  -6.2430  3.5053 -8.6055  5.2255 -6.2158
#>   0.0730  3.1052  5.9826  6.1585  2.1596
#>   3.4864  6.3010  2.8120 -12.0340  8.4295
#>  -6.8512  8.9314 -2.7023  5.6215 -3.1889
#> 
#> (1,5,.,.) = 
#>   1.3651 -12.0926 -6.4514 -7.8820 -1.1797
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]