Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   1.4943  6.9664 -1.4834  2.3634 -1.1946
#>   0.2530 -10.7428 -0.7497  4.4174 -5.1963
#>   0.8520  4.9596 -4.9931 -2.0989 -1.2978
#>  -3.4246 -1.1728 -1.9972  5.8548 -1.7924
#>   2.9862  7.1808 -0.3858  1.0861 -5.4370
#> 
#> (1,2,.,.) = 
#>   3.3272  2.9875  4.3631  4.1647 -8.7540
#>  -7.1812 -11.8856  3.3088 -0.8398 -3.9940
#>  -4.6105  0.2167 -6.9077 -6.1968 -5.3068
#>   2.5776 -6.5415  4.2555  7.4625 -0.5395
#>   5.3434 -4.6604  6.1015  2.6866 -2.2258
#> 
#> (1,3,.,.) = 
#>    4.6782   3.9170   9.4280   4.0653   1.3448
#>    5.3562  12.7732   9.0989   3.1316   4.7804
#>   -0.9213  -4.8293  -1.1718   2.9971   0.8567
#>    4.4256  -3.5157   9.4807 -11.6487  -5.7072
#>   -6.7220   1.5329   0.0373   4.3754  -2.0200
#> 
#> (1,4,.,.) = 
#>   -2.7142  -3.7479   2.4318  -1.3322   2.3338
#>   -5.0839   2.6362 -14.0296  -1.9564  12.4555
#>    9.6917  -1.8127   4.0119   3.8178   2.0427
#>   -1.0306   7.1102   4.9450  -2.0897   3.7443
#>   -0.2854   0.9138   3.6780  -2.1708   1.6550
#> 
#> (1,5,.,.) = 
#>    0.1751   2.8499  -1.3854  -1.5408  -1.0264
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]