Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    1.6143   5.2177  -5.1249  -2.5241   1.9136
#>   -2.1408  -8.3752   5.6519   0.8349   0.1847
#>   -3.9496   7.1213   5.5672   6.3373 -12.9238
#>    2.0896  14.2690 -11.8536  -8.2766  -3.7730
#>    2.5372  -0.2954  -7.1111  11.7987  10.7765
#> 
#> (1,2,.,.) = 
#>  -6.5370  2.0623 -0.5635 -4.1561 -1.4335
#>  -2.1260  0.1380  7.1199 -7.4534 -3.1369
#>   1.5048  2.0142  5.1431 -9.2787 -10.4867
#>   1.8227 -2.1797 -7.1387 -0.3190 -0.6468
#>   4.9882  2.3059 -0.6381 -0.6684  1.8046
#> 
#> (1,3,.,.) = 
#>  -0.7760  2.8608  0.1578  1.1610  1.8795
#>   0.5220 -0.6063 -8.1973  0.9941 -0.2200
#>  -1.2311  1.5557 -3.5172  1.7657  2.9316
#>  -2.0310  2.1547  1.2559 -5.0508 -3.1064
#>   0.8642  3.1143 -5.1636 -13.2596  6.1992
#> 
#> (1,4,.,.) = 
#>  -0.5398  3.4266 -4.1727 -3.8786  8.9245
#>   2.2193 -2.7921  7.6715  0.3154  6.2611
#>  -3.0677 -8.0089  9.9152 -5.5225 -4.8728
#>  -3.6713 -0.3605  6.3078  2.2428  0.8821
#>  -0.8825  3.8175  4.6798  2.8952  1.2478
#> 
#> (1,5,.,.) = 
#>    2.3167   0.4751  -2.4386  -2.6128   7.2562
#>   -0.5576  -5.7904  10.3523  -7.9593  13.2763
#>   -1.5729   3.8841   3.4354  -1.8210   1.2674
#>   -6.8549   1.5547   2.9867  -9.9613  -2.7272
#>    0.5849  -1.0630   4.4898  -2.7423   3.3727
#> 
#> (1,6,.,.) = 
#>   -2.0622  10.3492  -0.5201   4.3613   2.1982
#>   -1.7302   9.2725  -4.2618  -3.0880   6.7578
#>   -1.2928 -10.6017  -0.6298  -1.9082  -0.1840
#>    4.5269  -3.9393   1.7212  -3.3303  -6.7296
#>    0.9048   5.6542  -2.9283  -5.2215  -2.6986
#> 
#> (1,7,.,.) = 
#>  -5.1260  0.9345 -3.7561  6.5476 -0.4787
#>  -2.2023  5.6023  3.7178 -0.1647  2.5023
#>   2.0369  0.7183 -3.3544 -11.4650  1.9698
#>   3.7399 -8.6658 -1.6893  0.6574  0.7812
#>  -1.0749 -0.5485  4.9508 -0.5752 -3.1834
#> 
#> (1,8,.,.) = 
#>    0.5902  -2.2046   3.2423  -5.4686   3.9826
#>   -2.0733  -0.3782  11.3031  -3.4681   1.0814
#>   -3.2938  10.2905  -3.0860  -2.6057  -0.9223
#>   -8.2569  -0.8126  -4.3228   8.1291  -0.8599
#>   -0.7869  -3.3895  -3.8314   4.6396  -5.2937
#> [ CPUFloatType{1,8,5,5} ]