Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -3.6569  -0.5930  -0.6334 -11.4374  -1.1186
#>  -14.2296  -9.9521  -0.6080  -8.0338  -0.1871
#>   10.6689   4.5515   6.8241  -0.2414   3.5623
#>   11.1793   3.3345  10.4324   0.5486   3.2827
#>    0.5732   2.0520  -8.2761   8.7728   7.1496
#> 
#> (1,2,.,.) = 
#>   -0.0642  -6.9424  -5.6961   3.1038  -0.0831
#>    5.2847  -3.1092  -8.0795   0.3209   0.8021
#>   -0.5221   2.4501  -5.8932   0.2817   4.7333
#>    3.3829  -2.3245  21.1366   3.9237   7.4744
#>   -5.3289   2.8786  -0.6527   7.6367  -7.8615
#> 
#> (1,3,.,.) = 
#>   -2.9447  -1.7074  -8.1045  -4.2894   3.5820
#>   -7.6669   1.9077   6.5673  -2.4379   7.9067
#>    5.1050  -1.3660  -3.5154   1.3127  10.0989
#>   -2.3369   2.1310  19.0142  -2.8333  -3.6193
#>    1.9529   0.8538  -5.7973   5.0198   1.9899
#> 
#> (1,4,.,.) = 
#>    2.8779  -1.7491  -5.9069  -3.1991  -6.7325
#>   -7.6372  -9.3423  -3.9931  -2.4433  -0.9580
#>   -4.6517   5.4881   4.1319  -1.0373  -2.2326
#>    0.0554   5.3892  10.0705  -1.7087  -5.1660
#>    2.2235   0.5439   3.0222   0.6577   1.5862
#> 
#> (1,5,.,.) = 
#>    2.6939   2.2111  -3.4297   4.2696  -4.7281
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]