Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    1.7719   4.2754  -8.9503   6.7739   0.7566
#>    2.9198  -4.6655  12.9013   0.6139  -6.5159
#>   -4.9283  -6.3263  -5.0988   2.3228  -2.4844
#>   -5.9562  -2.7517   2.1690   2.9555  -6.6793
#>    1.1498   2.0101  -4.4752  -5.3457   4.8201
#> 
#> (1,2,.,.) = 
#>    2.4437   5.1564   3.9302 -10.1953   3.3291
#>    4.0725   1.1330  -6.1036  -0.2455  -5.4876
#>   -6.7839   1.4384  -6.3022  -1.0182  -2.2650
#>    1.7703  -7.4948  14.8109   4.3540  -4.8080
#>    4.8406   1.5108   3.8039   1.9475  -0.8510
#> 
#> (1,3,.,.) = 
#>    3.7223  10.1272  -4.3310  -4.1605  -1.7771
#>   -7.6397   5.1824  -3.5186   0.3484  -2.7837
#>   -0.4515  -0.2784  -3.7113   7.2957  -3.0204
#>   -1.4256   0.6799  11.0608   0.4379  -3.1782
#>    0.0048   3.1947  -4.2577   0.7694   3.1643
#> 
#> (1,4,.,.) = 
#>    2.9463  -3.6790  10.3054  -2.2362  -0.4402
#>   -7.2078 -11.0690  -5.9535   2.4936   1.1027
#>   -8.5147   0.0364   9.9520  -1.1500   1.8832
#>    5.9257   1.2203  -0.3514  -1.4728   6.2396
#>    6.3757   4.1044  -3.3067  -1.7531  -0.2602
#> 
#> (1,5,.,.) = 
#>  -6.3177 -0.5984 -0.5579  2.5351  0.3354
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]