Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -3.1170 -3.5056 -0.4593  1.0650 -1.9838
#>   1.7487 -7.2072  9.0959 -2.9232 -8.0376
#>   1.5660 -0.1774 -4.4523 -3.6660  3.6327
#>  -3.7712 -1.0107  2.5474 -0.0190  4.0275
#>   2.3844  4.8309  3.2915  4.4443 -0.4847
#> 
#> (1,2,.,.) = 
#>   2.3506  5.0022 -1.7100  0.7679 -8.7805
#>   0.9193 -5.1751  2.3088 -8.9268 -0.8132
#>  -3.3493  2.6482  0.7658  5.4330 -1.4544
#>   3.5011 -6.3058  2.4554  1.1977 -1.8167
#>  -3.4408 -6.1220 -1.7453 -0.0488 -0.6486
#> 
#> (1,3,.,.) = 
#>   4.8073e+00 -2.5342e+00  6.8365e+00 -4.8907e+00  3.5170e+00
#>   4.1485e+00 -1.3940e-01  1.0561e+00 -1.3293e+00 -9.9811e+00
#>   1.0571e+00 -1.1628e+00  7.9809e+00 -2.9758e-05  2.7629e+00
#>   3.5474e+00 -6.4913e+00 -1.8761e+00 -1.9076e+00  9.0632e-01
#>  -1.1616e+00  7.7559e+00 -8.2728e+00 -2.1109e+00 -2.4145e+00
#> 
#> (1,4,.,.) = 
#>   4.7686 -2.6410  0.1930 -2.1210  5.6976
#>  -0.2144  1.2087  0.0455 -18.5915 -1.2937
#>   3.8759  2.9834 -2.1451  1.1316  5.6533
#>  -1.7885  1.8732  4.4191  2.9819  0.9797
#>   1.3576 -7.3261 -7.1590 -2.1769  2.6675
#> 
#> (1,5,.,.) = 
#>  -3.8336 -1.3334 -2.8195 -6.1865  5.7207
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]