Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -0.3038 -0.7121 -5.0377  5.1648 -5.6998
#>   2.9595 -1.0588  2.0290  3.7307 -4.0111
#>   5.0668 -1.6883  0.4197  8.6468  1.6037
#>  -8.1864  6.3332 -3.2849 -2.7712 -3.3828
#>  -1.3831  6.2661  2.1227 -6.4798  2.3011
#> 
#> (1,2,.,.) = 
#>  -1.9469 -1.9192  3.6945  3.1902 -2.9006
#>  -4.3471  8.3144 -3.4242  5.2621  1.4652
#>   3.1440  2.2186 -2.4249  2.8892  3.3345
#>  -0.0477 -2.7622 -3.4909 -9.4254  2.4340
#>  -3.0337  4.8500 -6.5628  2.1934 -0.2526
#> 
#> (1,3,.,.) = 
#>  -2.7267  2.8472  0.0833  2.7238 -0.4757
#>   1.6508  6.7797 -0.8607 -2.2117 -4.0678
#>  -8.9883 -3.4002 -4.4214  0.5453 -2.4852
#>  -7.4940  8.2181 -0.7793  3.0940 -0.4733
#>  -1.5327 -0.2824  4.5294  5.5788 -4.6000
#> 
#> (1,4,.,.) = 
#>   -6.3762   8.3498   2.6012   2.0096  -5.0452
#>    6.2775  -3.8166   4.3948  -4.0905  -3.5371
#>   -0.6878  -8.2825   3.1387  -4.5314  11.1835
#>   -1.8779  -2.9619  -0.8188  -4.6212   3.1960
#>   -5.6537  -7.6852   4.9613  -2.7748  -2.8542
#> 
#> (1,5,.,.) = 
#>   6.6125 -4.8118  5.7719  5.1456  0.3453
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]