Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -3.8215 -14.8213 -1.6871 -10.0764  7.8492
#>  -1.5915  2.6677 -6.2840 -5.0990 -2.1940
#>  -6.8712 -13.5946 -3.1778  1.6937  3.1837
#>   1.7891 -4.6004  9.1279 -3.1999 -2.4716
#>   0.8199 -0.9711 -3.1153  5.9616 -0.2230
#> 
#> (1,2,.,.) = 
#>    3.4612  -5.0547  -1.0197  -5.3970  -0.0454
#>   -5.0291   1.4430 -10.0857   0.6451  -0.4315
#>    5.1932  -3.8831 -11.3792  -0.7616  10.1743
#>   -3.6750 -11.9477   1.6025  -5.9945   4.4840
#>    0.6405   4.7570  -7.6334  10.6526   7.5520
#> 
#> (1,3,.,.) = 
#>    3.9181   4.8076   8.3182   8.5035   1.2532
#>   -2.6941  -2.7226  -4.4715  -7.8970  -1.9171
#>    5.5729   6.5015  -5.6742  14.0539   5.3683
#>   -5.1991 -14.3230  -3.5960   1.4451   1.8565
#>    7.1690   8.3751  -1.8566  10.6876   5.9726
#> 
#> (1,4,.,.) = 
#>   -7.9848   1.4197   6.5887  -6.6406   3.0526
#>   11.5023  -2.4613 -10.5283   3.1548  -3.7721
#>   -6.5704  -1.7354  -2.1071  -8.6044   8.7651
#>    6.2342   0.9020   7.4567 -11.9605  -0.7255
#>   -4.2569   0.7786  -8.5757   4.3897  -2.2128
#> 
#> (1,5,.,.) = 
#>  -6.3794  4.8057  5.8071 -3.9317  0.5915
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]