Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>   7.8206  6.9397 -4.6352  2.7863 -0.0067
#>  -6.2321  1.2682 -1.3750 -2.6230  3.5049
#>  -5.4843  6.1034 -1.6059 -4.3805 -9.7593
#>  -3.6978  3.3345  6.5116 -10.8596  2.8146
#>   6.3172 -9.8955  3.0362  1.4565  1.8287
#> 
#> (1,2,.,.) = 
#>   -3.7301  -7.2624   1.4318  -5.0172  -4.7560
#>   -0.4734  -1.0971  -4.9234  -0.9646  -2.5377
#>    1.4878   2.6045   3.5234  11.0250   2.1951
#>   10.0509   6.3865   4.4260   2.2707   3.0506
#>   -5.5032  -7.3950  -3.0894  -5.5927  -0.6158
#> 
#> (1,3,.,.) = 
#>   2.3111  2.0930  1.9903 -2.3291  8.1578
#>   0.5850  4.1453  3.5996 -4.9863  3.2898
#>  -4.7333 -0.6774  2.4865 -6.2717  2.5820
#>  -3.7218  2.2452 -0.5671  1.1936 -6.2309
#>  -2.1697 -0.1201  2.0840  3.8001  5.4615
#> 
#> (1,4,.,.) = 
#>   3.0112  2.0374  2.3575 -5.1541 -0.6072
#>   1.5738 -1.3909 -1.6980  2.8539 -0.7350
#>  -3.4107 -12.5867  2.1012  0.2656  3.0777
#>   5.6777  3.8282  0.4708 -0.7511 -1.6607
#>  -2.2196 -6.0187 -9.4995 -1.1043  0.6854
#> 
#> (1,5,.,.) = 
#>   -1.6178   3.6163  -3.4003  -3.0144   3.7148
#>   -3.6156   4.7446  10.8532   0.0655   7.3331
#>    2.3652   7.8998   4.7579   3.5650   6.6269
#>  -13.6048  -4.1520  -2.1934   5.5759  -2.8188
#>    0.6818  -2.4957   6.1103  -1.5225   1.2129
#> 
#> (1,6,.,.) = 
#>   -0.6970  -4.4520   3.7204  -2.6246   0.3853
#>    4.5672  -7.2380   1.6334  12.9404  -4.6505
#>    8.6778  -3.3534   5.4396  -3.6520  -2.8669
#>    8.1110   4.6542   7.3615   4.3272   7.4934
#>   -1.8513   2.3617   0.0743  -9.7029   3.3382
#> 
#> (1,7,.,.) = 
#>    5.3701  -2.7725  -3.6187  13.8301   4.3158
#>   -3.9002   8.8952   2.6650  -8.0726   5.7906
#>    0.4209   9.9233  -5.1423   8.9969   5.5249
#>   -3.9430   0.4083   7.8620  -2.2777 -11.7264
#>    5.9327  -5.5501  14.2252  -5.1856  -2.5390
#> 
#> (1,8,.,.) = 
#>   -3.5848  -2.5115  -5.0085  -8.6458   0.2446
#>    4.6457  -3.5530   0.5775   1.6518   0.2303
#>   -3.5516  -1.2014   6.1369  10.7079   3.1213
#>   -7.6452  -1.2343  -5.4689  -0.0353   1.4719
#>   -1.5791  -4.0679  -6.2132   1.2951   2.7128
#> [ CPUFloatType{1,8,5,5} ]