Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    0.5219  -9.1593  -5.6033   0.7000  -1.4944
#>   -6.8091  -2.9887  10.5258   4.5513   1.9054
#>   -9.8785   1.0145 -20.8926 -10.1057   5.9239
#>    0.7028  -8.6833  -9.0341  -2.6711   3.4123
#>   -2.8483  -2.1543   4.7764  -3.3305  -2.2108
#> 
#> (1,2,.,.) = 
#>   -3.7319  -0.5349   6.8566   1.6935  -1.2933
#>   -1.8607  -4.0881 -10.6824   7.2440  -5.0744
#>   -5.1836  -9.8067   0.2800   4.3430   0.0993
#>   -4.4238  11.5000 -17.0532   1.7189   1.1326
#>    4.1377   3.9382   0.0671  -7.0371   0.7554
#> 
#> (1,3,.,.) = 
#>   1.0125 -5.9910  2.9450  1.5232 -6.0396
#>  -5.7652 -7.7419  8.8803 -3.2860 -7.8807
#>  -3.7029  7.1809  1.3045 -4.2012 -4.2485
#>   4.9834 -10.8008  1.5238  2.0593  0.3766
#>   1.9333 -3.3454  2.4484 -7.2243 -1.4140
#> 
#> (1,4,.,.) = 
#>  -0.4109  6.8162  2.5620 -6.7687  3.1419
#>  -6.7291 -0.5527 -1.7014 -0.7342  1.0597
#>  -4.8680  5.8581  1.8241 -9.3362  5.0554
#>  -6.5677 -1.9015  7.2538  2.3010  4.8108
#>  -1.4316 -6.9082  3.6513 -0.7764  0.8783
#> 
#> (1,5,.,.) = 
#>   6.3416  7.1123  1.6242  7.7547 -1.2812
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]