Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   -8.5973  -2.9768  -2.8793   6.3987  -5.4373
#>   -2.4548  11.1322   2.6035  10.8492   3.5734
#>   -3.1799 -13.7290  -8.0343  -5.8421  -4.2277
#>   -1.6060   7.0172   0.3022  -0.3428   7.7070
#>    1.3418   1.8720   0.7433  -3.2170  -5.5467
#> 
#> (1,2,.,.) = 
#>  -0.0415  3.1615  8.7689 -5.8346 -0.9504
#>  -14.3301  2.9458  3.6425 -3.0927  3.2360
#>   1.3115 -6.2359 -7.1189  2.4624  7.8361
#>  -2.4559  4.9882 -13.8497 -9.4355 -4.2715
#>   2.9064  5.7496  0.9894 -2.4231 -2.6849
#> 
#> (1,3,.,.) = 
#>   2.3956 -1.0348 -0.8075  0.8657  3.9583
#>   3.5240 -1.3498 -8.6479 -4.0444 -2.0519
#>   1.4662  2.0147  6.2455  9.3604 -0.3796
#>   5.2707 -9.7525 -5.6014  1.6384 -0.2863
#>   2.1549 -2.7809 -8.0806  4.8323  3.8186
#> 
#> (1,4,.,.) = 
#>  -0.1812 -3.8493 -4.4492  9.5919 -0.6385
#>   2.7168 -1.5496 -5.0477  1.5821  4.1259
#>  -2.5032 -9.2154  2.7888 -3.7239  5.9721
#>  -0.7291  3.5664 -1.0759 -3.1059  7.6140
#>  -0.2098 -0.1563  0.9866  2.7631 -9.5222
#> 
#> (1,5,.,.) = 
#>   5.7716 -6.5641 -0.7910  7.3051 -3.3144
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]