Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   5.0376  2.5287 -2.5290  6.5732  0.9235
#>   7.3123 -8.2556 -7.2983  1.5073 -5.0618
#>  -1.0148  5.0664 -1.5813 -6.9495  1.7552
#>   1.9476 -8.0841  3.1500 -10.7651  1.8977
#>   7.1355  3.1359 -1.8356 -1.7660 -2.0481
#> 
#> (1,2,.,.) = 
#>  -3.4460 -8.6309  2.2088 -1.3413 -2.3101
#>   5.0521 -6.7978 -1.8494  5.8924 -6.9346
#>  -3.6800 -2.8273  0.0374  8.5391  4.8661
#>   0.7432 -7.0999 -1.7963 -0.7237  3.4861
#>   2.9596 -2.3666 -6.9526 -1.5871  3.5109
#> 
#> (1,3,.,.) = 
#>   5.0232  4.2504 -11.1130  2.3826  0.7471
#>   0.5773  0.1487 -16.8129 -3.2094 -1.0799
#>   5.9381 -4.3120 -8.3686 -3.3504  0.5184
#>   1.1966  2.8506 -0.3846 -0.7312 -1.5923
#>  -0.6166  0.4632 -6.7537 -4.1985  2.4502
#> 
#> (1,4,.,.) = 
#>    8.5649   0.5255  -2.9174   5.8590   0.7771
#>    2.4931  10.2844  -7.8331  -7.6508  -3.8648
#>    1.8777  -4.9012  -1.4303  -2.9161   6.0661
#>    3.5730   0.1929  -6.8675  -8.9981   1.1431
#>    3.7159   4.8578  -0.0442   1.7795  -1.1776
#> 
#> (1,5,.,.) = 
#>  -6.3995 -1.3425  6.3824 -2.1979 -3.8260
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]