Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>   -4.5588  -8.3236  -6.4814   4.3252   4.2117
#>    0.7229  -0.6559  -7.4490   6.5644   0.6572
#>    3.8078   0.2284  -9.0923  -9.2898   9.0271
#>    0.7561  -2.1838   1.0426   8.2051   3.3293
#>   -2.6024   5.5291  -2.8092  14.0497  -0.2068
#> 
#> (1,2,.,.) = 
#>    0.4685  -8.3725   0.6208  -2.8500  -1.2229
#>   11.2077  -2.8395  -8.0856   4.8323   5.0084
#>    0.8266   6.4707  -5.9195  -1.1153   5.3364
#>    2.8211  -2.4830   6.3360 -12.7959   3.6628
#>    0.0738   8.9628  -3.6424  -0.4259  -5.0392
#> 
#> (1,3,.,.) = 
#>  -2.5726  0.7886 -4.3724 -3.5194  6.0941
#>  -3.9185 -1.1170 -3.0384  1.3922  1.3140
#>   3.2885 -7.4759 -2.7243  8.6262 -2.4189
#>   3.7545  4.5061 -1.7533  5.8961  0.3540
#>   0.4908 -7.3335 -3.6133  0.0636  3.7707
#> 
#> (1,4,.,.) = 
#>   -2.3616  -0.5667   1.5306  -0.6928   0.4921
#>    0.7869   7.7548   7.9711 -11.6979  -1.8266
#>   -0.9844  -2.1805   0.4782   8.6756  -6.7048
#>   -3.4391  12.1285   3.9507 -25.9902   7.5459
#>   -0.3228   0.2769   0.0451   2.0257   4.3603
#> 
#> (1,5,.,.) = 
#>   1.3584 -6.2070 -0.8751 -6.2547 -1.0638
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]