Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -8.1985  1.2517  1.5634  1.7102  2.7358
#>  -9.3066 -2.0839 -3.1107 -5.0281  5.8559
#>  -0.2180 -2.6929 -2.6181  9.8450 -4.0013
#>   6.4429  0.1812 -5.0130  4.2632 -0.6818
#>   2.7767 -10.3776  2.4959  0.9270 -3.3845
#> 
#> (1,2,.,.) = 
#>   -5.0719   3.5521   0.4529  -4.4392   7.3479
#>   -1.3189  -0.6911 -12.6956   0.9782  -2.5182
#>    5.3132   5.7202  -7.4993   0.3517   2.5389
#>    4.5631  -5.5485  10.3901  -2.1856   5.2610
#>   -0.5512   2.3227  -3.6441  -3.7616  -0.5447
#> 
#> (1,3,.,.) = 
#>    3.4874  -6.2364  -0.7306  -2.1287   2.9625
#>   -3.4671   5.8528  -3.0658   1.7257  -5.4566
#>   -8.5196   2.4360  -2.2770   7.3017   7.7905
#>    1.3967   4.5635   4.1271  -6.3795  10.5440
#>   -0.0962   0.0544  -2.7040  -4.5356   1.5189
#> 
#> (1,4,.,.) = 
#>    0.2679  -4.7315  -0.5954   2.2860   1.0403
#>    1.9219  -4.2472  -5.8120  -0.3949  -4.7339
#>    4.8081  11.6706   2.8379   0.6288   0.0120
#>   -5.8532   0.5320  -3.0480   3.4904  -2.1287
#>   -5.8261  -1.0541  -2.9553  -1.4941  -1.8812
#> 
#> (1,5,.,.) = 
#>   1.2721 -9.5863 -9.3915  1.3882  2.1539
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]