Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   10.4014   0.2968  -2.1756 -13.8829  -4.5546
#>    3.9839   6.5896  10.3713  -0.1200   4.0398
#>   -6.3200   0.1209 -15.9031  -8.1899   6.5320
#>    0.9248   2.4629   2.1555   6.6168  -5.3858
#>   -2.5836   5.1039   1.2085  -5.4472  -2.7499
#> 
#> (1,2,.,.) = 
#>    6.2136   6.4077  -2.4482  -5.8519  -4.4928
#>    7.5653   6.9908  11.5249  -3.7301  -9.6363
#>   -2.2693  -0.6840  -4.6371  -5.5631   7.6900
#>   -1.1067   2.0523   2.0983   7.0708  -6.7416
#>    0.3457  11.4241  -2.0024  -2.9042   1.7810
#> 
#> (1,3,.,.) = 
#>   -2.5403   3.2356  -4.4038  -3.8422  -0.8787
#>    9.8354   2.2426  10.5683  10.1336   5.7692
#>   -1.9301   1.0621 -10.7798  -7.0996  -4.0512
#>    3.1655   3.6425  -5.2887   0.4786   1.9588
#>    5.3202  -3.7369  16.0627  -4.7859  -4.0363
#> 
#> (1,4,.,.) = 
#>   -2.9422  -3.9192   0.1029  -7.2455   1.8441
#>    4.9016  -2.2553   5.8698   2.6909  -1.6923
#>   -8.6711  -7.8781   1.2312   4.5118   4.5077
#>   -8.0919  10.1968  -2.6644  -3.5213   7.9337
#>    6.1006   5.3779  -2.5477   2.5329   0.7578
#> 
#> (1,5,.,.) = 
#>  -10.0465   2.6384 -13.2687  -4.7452 -12.5834
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]