Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    0.6443  -2.8581  -3.8018  -0.7160   2.0964
#>    2.3809   2.0079  -4.5665  -0.5414   0.2336
#>   -2.7114   1.2678  -2.9845  -6.0529   3.9594
#>   -6.9777  -1.4925   0.1524   4.9399 -10.4250
#>    8.7126  -5.6973  -5.4110  18.1268   2.7683
#> 
#> (1,2,.,.) = 
#>    5.2218   3.1188  -8.6960   5.9075   0.9253
#>    3.7407  -0.1224   7.1301  12.9983   2.1768
#>   -6.7623  -8.0547   5.0581   3.4517  -6.4046
#>    3.2942  -8.5740   4.4801  12.5073  -2.2104
#>   -3.4364  -0.2698  -2.0076  -9.5942   1.7210
#> 
#> (1,3,.,.) = 
#>   -5.2061   3.5831   2.1301   2.9067   3.7096
#>   -1.9463  18.2973  13.3650  -6.8584  -1.0309
#>    2.2064   2.1724  -1.6561  -3.6185  -5.1734
#>    1.5636   9.7497 -11.2415   3.0867   0.7855
#>    4.1923  -4.8290  -2.6237   2.4276   3.9101
#> 
#> (1,4,.,.) = 
#>  -2.3121 -7.9832  0.8913 -4.1757 -2.4864
#>  -0.5757 -7.8273  7.4204 -2.6988  2.7922
#>   0.9479  1.9199 -3.1743 -3.1409 -1.5770
#>   1.3446 -6.5043 -8.6385 -9.4068 -4.4736
#>  -1.3072 -3.7244 -0.1214 -0.8011  0.2729
#> 
#> (1,5,.,.) = 
#>   0.7057 -1.1460  0.2666  0.0249  2.8535
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]