Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>  -7.5110  3.7641 -10.4372  4.2887 -1.1483
#>   2.5014  4.1213 -1.3290  2.7715 -1.2356
#>   3.9241  2.2920  3.0992 -5.0194 -3.9354
#>   1.2425  5.6448  5.1804 -11.2531  3.6939
#>   2.7946 -0.2876 -1.8022 -2.9866  1.7587
#> 
#> (1,2,.,.) = 
#>  -5.7366 -1.2901  8.7096  0.2504 -3.9261
#>   3.7174 -1.6981  8.2566  1.5653  2.0469
#>   8.4169  8.5922  4.2183  3.2537 -8.8162
#>  -0.8450 -4.8588 -14.4069 -2.3530  4.7965
#>  -2.7730 -0.7138 -3.8965  3.0225  3.0301
#> 
#> (1,3,.,.) = 
#>   1.7159  4.1838 -1.0616 -0.8207 -5.0284
#>  -6.3134  2.0706 -3.1550  7.4245 -1.5805
#>   5.3365 -5.3354 -5.2464 -8.8019  7.0449
#>   6.8231  3.0109  6.6110 -3.6602  3.4779
#>  -0.5472  2.4284 -5.0799 -4.8086  3.5136
#> 
#> (1,4,.,.) = 
#>  -1.0399  4.2911 -1.7679  4.1818 -3.6969
#>   1.6923  3.2699 -2.3755 -2.7303 -6.1423
#>  -1.7567 -0.9807  2.1321  6.2346  3.4545
#>  -2.7856  1.0777 -0.1440  3.6303 -4.2952
#>   0.8700 -4.6312  0.7110  0.2863 -0.3070
#> 
#> (1,5,.,.) = 
#>   6.3820 -3.1108 -3.1401  1.8361  1.4169
#>  -3.5173 -1.9328 -3.2543  5.3292 -0.7929
#>  -5.1491  4.1741  1.0536  7.7839 -1.0521
#>  -0.5027 -0.5993  0.8675  3.5631 -2.8527
#>  -2.5527 -6.5371  5.4238  6.4524  3.9158
#> 
#> (1,6,.,.) = 
#>   0.1192  0.7898  5.1046  0.8291  0.8075
#>  -0.1448 -0.4707  4.0276 -6.0713  0.6670
#>   3.9042 -0.1029  0.2932 -0.5353  4.6275
#>   0.2660  5.2588 -5.4768 -5.2114 -6.2913
#>   1.3878  3.3422 -7.8499  2.2837 -1.2931
#> 
#> (1,7,.,.) = 
#>   0.3207  1.6438 -0.1820  2.9592 -5.9365
#>  -1.8610 -1.4471  2.4745  4.9803 -2.3983
#>   2.1440 -2.1082 -3.0073  4.2901  8.0834
#>  -2.2489  1.8009  1.5683  8.1625  2.1253
#>  -3.1867  0.6020 -0.8497 -3.0459  3.1883
#> 
#> (1,8,.,.) = 
#>   -4.5885   1.7673 -10.5134  10.8415   1.1800
#>    3.8736   5.2990   0.8102   3.4449   3.6823
#>   -2.2823  -1.8588   6.1128   4.2038   2.2027
#>   -0.3619   4.1382   3.2630   0.0729  -8.0006
#>   -0.2060  -0.7592   4.9095  -2.3508  -7.0914
#> [ CPUFloatType{1,8,5,5} ]