Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor 
#> (1,1,.,.) = 
#>    5.6427   2.6900  -0.0572   3.6534  -2.4089
#>   -2.2426  -1.3052  -1.0697  -4.1453   3.4645
#>   -3.1559  -7.4690  -6.5846   5.2213   1.3986
#>    1.8330   5.0599  -2.4154  17.5550   3.0685
#>   -1.2203   1.9489  -5.5997  -5.3750  -0.3962
#> 
#> (1,2,.,.) = 
#>    2.4205  -6.9905  11.4627   2.7888   1.3995
#>   -4.0647   8.9093   2.3628  -0.0779  -1.4396
#>    6.7686   1.4683  -2.5210  -8.7289   2.8817
#>    6.7092  -4.5630   2.8886   3.2551   8.3960
#>    1.0024  -0.3961  11.9043   8.5755  -2.2479
#> 
#> (1,3,.,.) = 
#>   1.1439  2.3956  2.1450 -3.4417 -3.3577
#>   6.3678 -3.0977 -2.8115  3.0435  0.0480
#>   0.9487 -9.6482 -4.3639  0.8018 -6.0693
#>   7.2841 -2.9704  1.5337  3.9896  7.5048
#>   8.6867 -1.6283 -6.8413  7.6310  0.0267
#> 
#> (1,4,.,.) = 
#>   3.3795  2.6372  6.3721  1.4735  1.1823
#>   2.2618  3.8031 -3.6062 -3.4912  5.9428
#>   0.8257 -5.1489 -1.4126  6.3720 -2.4351
#>  -0.2114  4.3334  8.8164  4.8549 -2.1184
#>   2.9252 -2.3333 -6.6579  0.6540 -1.2040
#> 
#> (1,5,.,.) = 
#>   1.3367 -3.7454  0.9575  1.0511  1.4883
#>   0.5365  1.5529 -4.4741  3.2148  0.9415
#>  -2.2668 -2.6881 -0.4441  2.3668 -6.7160
#>  -1.2787 -0.3678 -4.2080 -5.7846  1.1154
#>  -1.9345  1.8609 -12.9168 -2.0863  2.6221
#> 
#> (1,6,.,.) = 
#>   -1.0741   1.1927  -2.4475  -3.7062   2.3783
#>    0.0469   0.1178  -1.2013  -0.2609   3.9160
#>   -1.9084  -5.8541   3.7067   8.9891  -2.3641
#>  -10.7526  15.8941   2.4833  -7.3412   1.7073
#>   -3.2317   3.9390   4.9352  -8.2996   4.2148
#> 
#> (1,7,.,.) = 
#>    0.3704   4.4950   6.0446  -4.1027   4.5159
#>    0.9657   0.9625   8.0026  -3.4499  -0.6313
#>   -9.1112   2.5262  -9.6846  -0.9841  -1.8033
#>    3.2470  -1.1839  17.3548  -8.0363  13.1482
#>    9.7550  10.8305  -0.1344  -1.3703  -6.1479
#> 
#> (1,8,.,.) = 
#>    0.0393   5.5237  -0.0691   4.1614   2.6568
#>    6.6200  -3.5911  10.1358  -8.9759   3.5868
#>   -1.3139   5.3733   0.5546   1.2136  -4.8962
#>   -3.6796  11.3923  -0.6634  -8.9935   7.8473
#>   -0.1899   5.6305   1.5438  -7.0192   0.8536
#> [ CPUFloatType{1,8,5,5} ]