Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>  -2.5804  5.2382  6.4005 -3.4090 -3.2509
#>  -14.2546  3.3503  6.1268 -2.1070  7.5787
#>   1.1690  7.0194  2.0886  4.6178 -1.8686
#>   4.8466 -2.7784 -1.1550  4.3557 -7.4315
#>  -2.5500  3.0978 -3.4047 -3.5223  2.8951
#> 
#> (1,2,.,.) = 
#>   0.8109 -9.3041  0.1802  7.0184  3.7951
#>   2.4880 -3.6727  3.9862  1.4844  4.6036
#>   9.6268  6.0975  0.9000 -11.3110  3.9963
#>  -8.3058 -1.4628 -7.6014 -2.4886 -2.5528
#>   1.5489  2.9587  1.4988  4.9042 -2.8225
#> 
#> (1,3,.,.) = 
#>   0.5246  5.7951 -6.1786  9.2342  4.4156
#>  -2.2840  6.8055  0.5748 -4.2639 -0.2779
#>   2.4357 -0.2987  4.4529 -10.9028  2.5270
#>  -4.1993 -0.2116 -3.2950  6.1035 -1.9086
#>   6.5157  1.3523  6.0694  3.2390 -0.0934
#> 
#> (1,4,.,.) = 
#>   3.4232  6.2227 -0.1714  4.3522 -3.4448
#>  -2.7954 -8.7819 -2.9668  7.0185  8.9268
#>   1.0682 -2.9950  2.3681 -2.6749  6.5567
#>   1.0554 -0.0929  2.9866  0.4770 -0.8907
#>   3.4972 -5.5135  2.6541  0.3074 -1.7143
#> 
#> (1,5,.,.) = 
#>   1.3756 -3.4109 -4.8709 -0.3270 -0.9105
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]