Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input
input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight
filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias
optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride
the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding
dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding
additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups
split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation
the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}

#> torch_tensor
#> (1,1,.,.) = 
#>    1.8842   7.0953   1.5857  -5.3931   1.5218
#>    1.4429 -15.9399   1.8437   2.6725  -0.1393
#>   -0.9690  -4.7004  -0.9740  -1.0323  -5.8271
#>    0.3488   1.9825  10.9060  -4.0180  -7.4521
#>    4.1293   8.4119 -10.8184  -1.8201   2.0118
#> 
#> (1,2,.,.) = 
#>   -1.8275   0.0753   0.0942   2.0565  -0.8016
#>    2.2925  -9.0406   3.4011  -0.8929  -2.1858
#>   10.6648  -2.7973  -7.2874   9.4497  -4.6176
#>    2.0520  -4.8590   8.7348  -9.0626  -3.3114
#>    6.1919  -2.2804  -3.4680   7.8420  -7.1956
#> 
#> (1,3,.,.) = 
#>    1.7832  -1.0122  -5.1522  -0.1208  -2.6428
#>   -4.9514  11.1895  -9.5480   2.8571   0.7190
#>   -3.6965   0.4481   1.1401   4.8694   0.8347
#>    2.3847  -2.3975  -1.9349   2.3427   0.9021
#>    1.6721   6.0271  -4.8243  -2.3756   0.5617
#> 
#> (1,4,.,.) = 
#>   -5.5310   8.6625   2.0705   6.8262  -0.2973
#>  -17.2414  16.9928  -8.6010   3.8661   0.9809
#>   -2.4461   3.4826  -7.3570   5.8736  -4.3387
#>    3.0010  -3.6378   6.6345  -2.4932   9.8843
#>   -4.0312   0.4255  -7.2313  13.0355   0.3129
#> 
#> (1,5,.,.) = 
#>   7.0501  3.6507 -1.4565  0.1574 -3.1879
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]