Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    2.8483  -7.6956  -5.1836  -5.1682   3.4370
#>    1.5629  -0.7483  -0.2548  10.6365  -3.3916
#>    3.0039  -4.2351   5.5557  -1.3791  -2.0405
#>    2.2593  -2.1874   6.4652  -9.8284   3.3464
#>   -1.8374   1.1539  -0.7373  -2.7115  -2.4165
#> 
#> (1,2,.,.) = 
#>  -3.7051 -4.7887  3.2116 -0.9423 -5.8319
#>   2.3530 -0.4831  1.6626  6.0272  7.3081
#>   7.0124  0.2066 -7.1514  1.6255  3.4127
#>   1.3425 -6.5303 -1.7103  5.5466  5.2975
#>   2.9455 -0.9673 -3.3204 -1.1437 -0.6503
#> 
#> (1,3,.,.) = 
#>    2.4419  -0.6673  -5.5785   0.3795   2.0246
#>   -4.2175   1.9549  -0.3164   0.9094  -6.9215
#>   -2.0359  14.5588  10.7931  -3.1418  -1.9947
#>   -1.6801   7.7935  -4.9022 -19.5539   2.9221
#>    1.6270  -6.5687   1.0141  -2.4670   2.0618
#> 
#> (1,4,.,.) = 
#>   -5.5697  -2.3566  -2.4726   1.5613  -1.7863
#>   10.2919   3.6186  11.5075   1.5992  -2.2791
#>   -9.8564  -0.3314  -4.1390  -2.3287   1.6227
#>    1.4585  14.3384  -6.6097   1.0276   3.0301
#>    0.2635  -6.1865  -7.3291   3.4226  -2.4468
#> 
#> (1,5,.,.) = 
#>   -4.7329  -0.6871  -3.4076  -3.4421  -3.8846
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]