Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   1.4904  1.1201 -3.5911  2.2584 -3.2491
#>  -4.1465 -10.3456  1.5974 -3.7265 -1.8778
#>  -4.8975  5.3252  6.7727  1.9358  2.3765
#>   6.1434 -2.3754  1.9620 -7.0355 -1.4174
#>   1.1557 -0.8196 -2.1665 -2.0107 -0.9631
#> 
#> (1,2,.,.) = 
#>   -4.5450  -2.2196   2.8124  -5.0696  -4.0279
#>    0.0605   7.7477  -3.7784  -0.0444   0.1461
#>   13.7037   3.8826 -14.8449  10.1081   0.1030
#>   -0.9413   3.6606   4.4162   2.8410  -1.8650
#>   -6.2069  -0.3486   4.6446  -7.9360  -0.1316
#> 
#> (1,3,.,.) = 
#>  -0.2236  3.9752 -6.4824 -1.5796  7.4289
#>   4.9271 -2.2612  3.5011 -3.8106  4.2970
#>   0.2663  3.0168  5.5375 -4.6676  6.9932
#>   8.3015  4.5517 -0.9761 -1.2982 -2.7744
#>   3.2358  4.6001  1.1790 -5.9229  4.8956
#> 
#> (1,4,.,.) = 
#>   1.7945 -7.0564  6.0941  3.5956  2.5207
#>   0.1439  3.1053 -5.1568 -4.8034  3.3033
#>  -13.5532 -0.3154  0.3022 -0.3512 -4.7462
#>  -4.6054 -5.1192 -0.1881  0.9681  0.0130
#>   7.0522 -3.7483 -2.2034 -3.1064 -0.1584
#> 
#> (1,5,.,.) = 
#>   -1.9007  -2.8228  -3.7172  -2.0452  -4.4214
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]