Conv_transpose2d

Conv_transpose2d

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>   1.3273 -0.4437 -0.5628  1.4033 -2.5892
#>   2.8951 -3.3222  1.7303 -11.4635  2.3463
#>   0.7087 -4.1212  3.7215  9.3307 -5.8969
#>   0.5380 -3.8452  1.9286  7.9421 -3.8007
#>   0.1344  3.9868  2.6462 -0.4598  2.5415
#> 
#> (1,2,.,.) = 
#>  -1.9967 -7.4634  3.1238  0.9362  3.0009
#>  -0.3406  0.9657 -1.5912  4.0337 -1.8880
#>   0.2630  0.3022 -8.0036 -3.4944  4.3138
#>  -9.3480  2.7566  4.6362 -9.1396  7.0318
#>   2.3405  1.0081  0.6607  3.0178 -4.8881
#> 
#> (1,3,.,.) = 
#>  -11.3032  -2.1097  -0.7187   4.2765  -4.3086
#>    3.6640  -1.1209   0.9784  -6.3973  12.1373
#>   -1.9044   5.4241   0.0494   5.4419   0.4567
#>   -0.8646   8.6790   0.0863   1.9881  -5.0270
#>   -3.4620   2.0984   1.4710  -1.7336   4.4210
#> 
#> (1,4,.,.) = 
#>    1.5779   0.1307   0.9466   0.6460  -0.4305
#>   -3.0424  -0.9121   0.3894  -2.8021  -3.2869
#>   -5.1986   0.5681  -1.3269  -2.4607  -2.4514
#>   -1.8291  11.9320   0.3594   4.5611  -2.2734
#>    3.5819   8.1434   2.4657   1.2124   0.5711
#> 
#> (1,5,.,.) = 
#>   -5.1348   0.2950  -3.2065   7.1313   5.3910
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]