Conv_transpose2d

Conv_transpose2d

Usage

torch_conv_transpose2d(
  input,
  weight,
  bias = list(),
  stride = 1L,
  padding = 0L,
  output_padding = 0L,
  groups = 1L,
  dilation = 1L
)

Arguments

input

input tensor of shape \((\mbox{minibatch} , \mbox{in\_channels} , iH , iW)\)

weight

filters of shape \((\mbox{in\_channels} , \frac{\mbox{out\_channels}}{\mbox{groups}} , kH , kW)\)

bias

optional bias of shape \((\mbox{out\_channels})\). Default: NULL

stride

the stride of the convolving kernel. Can be a single number or a tuple (sH, sW). Default: 1

padding

dilation * (kernel_size - 1) - padding zero-padding will be added to both sides of each dimension in the input. Can be a single number or a tuple (padH, padW). Default: 0

output_padding

additional size added to one side of each dimension in the output shape. Can be a single number or a tuple (out_padH, out_padW). Default: 0

groups

split input into groups, \(\mbox{in\_channels}\) should be divisible by the number of groups. Default: 1

dilation

the spacing between kernel elements. Can be a single number or a tuple (dH, dW). Default: 1

conv_transpose2d(input, weight, bias=NULL, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor

Applies a 2D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution".

See nn_conv_transpose2d() for details and output shape.

Examples

if (torch_is_installed()) {

# With square kernels and equal stride
inputs = torch_randn(c(1, 4, 5, 5))
weights = torch_randn(c(4, 8, 3, 3))
nnf_conv_transpose2d(inputs, weights, padding=1)
}
#> torch_tensor
#> (1,1,.,.) = 
#>    2.8477  -2.1225   3.4169   5.6890  -5.8392
#>    3.0017   3.5697  -6.1720   2.2051  -0.8895
#>   -3.9990  -8.8474   4.6073  -1.9908   1.9152
#>    1.9764  -3.1568   1.0030   0.4302  -2.0830
#>   -9.2667  15.1049  -6.9380   1.1301  -2.9126
#> 
#> (1,2,.,.) = 
#>  -5.2478 -1.8675  5.1962 -2.9033  3.9666
#>  -4.6268 -12.5972  3.5733 -8.6220 -4.0030
#>  -8.6305 -6.3559  5.5231 -2.1144 -2.2783
#>  -3.6451  8.8870  1.1079 -11.4230 -10.3722
#>   0.3594  0.3009 -4.4675 -15.4113  1.6759
#> 
#> (1,3,.,.) = 
#>    4.4084  -6.2714   1.6358 -10.2975  -1.6795
#>    0.1165  -0.5902   0.9830  -7.8685  -2.1431
#>    1.0183   1.9084   3.2124   0.0176   1.8053
#>    3.6783   6.0380 -10.9509  11.2613   6.9808
#>   -3.2693   2.2762  15.9107  -3.4232  -4.5632
#> 
#> (1,4,.,.) = 
#>    5.2747   1.7523  -9.6912   0.3707   3.6440
#>    4.0845   1.3434   3.1775  -9.6013  -1.5701
#>    3.3613   2.6192  -9.1587  13.5577   5.3618
#>   -3.5439   3.0355   8.4852   7.3001  -6.4887
#>    7.3604   3.8284  -0.6204   1.7107  -2.5265
#> 
#> (1,5,.,.) = 
#>    1.3481  -3.0242  -5.1534  -6.5716   4.1177
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{1,8,5,5} ]