Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   3.1269  0.9525
#>  -2.5551 -9.2671
#>  -0.3517 -3.9285
#> 
#> (2,.,.) = 
#>   1.3474 -1.2334
#>  -4.5162  4.7339
#>   0.2207 -1.8316
#> 
#> (3,.,.) = 
#>   4.1118  2.3770
#>   4.5216  0.1284
#>   2.5540 -10.7896
#> 
#> (4,.,.) = 
#>   2.4745  5.5189
#>   4.1149  2.1041
#>  -2.7162 -7.8699
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>   2.7652  1.9038
#>   0.3913 -0.5752
#>  -0.0733 -6.1049
#> 
#> (2,.,.) = 
#>   1.4418 -0.6379
#>  -2.4266 -4.5066
#>  -2.2360  2.4495
#> 
#> (3,.,.) = 
#>   0.8542 -0.2390
#>   0.5920 -3.9942
#>   1.1744 -1.2541
#> 
#> (4,.,.) = 
#>  -1.9343 -0.0744
#>  -1.1117 -0.1911
#>   0.7832  0.9810
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   -3.0607  -2.0808
#>   -0.7614  11.3993
#>   -9.7078  -1.1342
#>    5.3198  -1.8183
#>   -2.2561   1.1860
#> 
#> (2,1,1,.,.) = 
#>   5.9581  5.6764
#>  -0.7409  5.7791
#>  -2.2354 -3.3075
#>   1.5357 -5.2179
#>  -0.4934 -10.6985
#> 
#> (3,1,1,.,.) = 
#>   5.6713 -1.4706
#>   7.7647  0.3945
#>   1.7536 -2.3804
#>  -2.7693 -1.4015
#>   7.3998  7.8926
#> 
#> (1,2,1,.,.) = 
#>  -7.3846  1.0945
#>   2.6440 -0.0749
#>   0.8468  0.6714
#>   2.8560 -0.5236
#>   4.6443 -11.6895
#> 
#> (2,2,1,.,.) = 
#>  -2.6286  4.1849
#>   0.0714  9.0369
#>   2.2322  3.0162
#>  -6.9693  3.6747
#>   2.1045  5.4243
#> 
#> (3,2,1,.,.) = 
#>   12.0462  -0.1237
#>    6.1222   4.7002
#>    2.4700   9.5251
#>   -2.4001   6.1526
#>    6.2922  -4.2304
#> 
#> (1,3,1,.,.) = 
#>   7.1441  1.4981
#>  -4.1623 -0.3252
#>   2.9155 -1.5450
#>   7.9392 -1.9687
#>  -2.3094  0.0935
#> 
#> (2,3,1,.,.) = 
#>  -0.9630 -6.0835
#>   0.4546 -6.4689
#>  -0.0282 -9.4388
#>  -7.4903 -1.5077
#>   1.7163 -6.8045
#> 
#> (3,3,1,.,.) = 
#>  -8.3626  2.2797
#>   3.1811 -3.3831
#>   4.2462  2.6255
#>  -3.0474  6.3158
#>   6.0139  3.5483
#> 
#> (1,1,2,.,.) = 
#>   1.6137  2.7773
#>   0.8053  2.4957
#>   2.0098  4.0531
#>   7.1536 -4.0740
#>   0.5047 -3.4478
#> 
#> (2,1,2,.,.) = 
#>    6.9573  15.8894
#>   -2.4017  -1.8492
#>    6.6191  -0.4522
#>    0.2472   0.4819
#>    2.7057   4.7820
#> 
#> (3,1,2,.,.) = 
#>  -5.5708 -0.4651
#>  -3.3221 -7.2540
#>   4.7071  8.4420
#>  -1.0274  5.4142
#>  -6.8729 -0.6304
#> 
#> (1,2,2,.,.) = 
#>   1.1419  4.0883
#>  -4.2120  4.7751
#>  -0.8735 -2.5643
#>  -2.3289  0.6712
#>  -3.3304  4.8199
#> 
#> (2,2,2,.,.) = 
#>  -0.5676  9.9439
#>  -3.4647  4.7481
#>  -1.3951  0.8998
#>   5.0574 -4.5266
#>  -6.7789 -1.1194
#> 
#> (3,2,2,.,.) = 
#>   3.8036  6.1886
#>  -7.1878  3.4328
#>   1.1591  3.7105
#>  -7.5438  0.0551
#>   8.2602  9.0398
#> 
#> (1,3,2,.,.) = 
#>  -5.2171 -13.2233
#>  -2.0567 -1.1374
#>   4.2039 -1.5208
#>   1.9455 -4.7267
#>   9.0767 -4.1536
#> 
#> (2,3,2,.,.) = 
#>   3.7466 -4.5032
#>   1.4704 -2.0007
#>  -1.9988 -0.1092
#>  -4.4596 -0.3324
#>   2.1152 -13.2775
#> 
#> (3,3,2,.,.) = 
#>   -3.3043   0.5113
#>   -2.8777   4.2746
#>   -3.3955   1.7092
#>    9.2973  10.1919
#>   -1.5666   2.8480
#> 
#> (1,1,3,.,.) = 
#>    0.3320   4.6616
#>    0.8167   2.9600
#>    0.4455   4.4140
#>   -1.3307  10.8842
#>   -9.4580   1.1702
#> 
#> (2,1,3,.,.) = 
#>   -1.8528   7.6680
#>    8.2839  14.2152
#>    2.9796  -3.5969
#>   -1.7270   1.0632
#>    0.1975  -0.5567
#> 
#> (3,1,3,.,.) = 
#>  -0.7397  2.4980
#>  -2.3425  1.0752
#>  -1.2656 -9.2947
#>  -7.2362 -3.9018
#>   5.9827  2.7699
#> 
#> (1,2,3,.,.) = 
#>  -15.5498   1.0546
#>    1.3051   0.4347
#>   -1.4093   7.5153
#>   -6.1192   4.7335
#>   -4.9058  -4.5730
#> 
#> (2,2,3,.,.) = 
#>  -3.6604  1.9234
#>  -0.8458 -0.4348
#>  -1.8828  3.6361
#>   1.8287 -1.8895
#>   2.6902 -5.2685
#> 
#> (3,2,3,.,.) = 
#>  -1.3802 -0.7002
#>   2.3379  2.1387
#>   0.3050 -2.9317
#>   8.0050 -0.7583
#>  -5.9590 -1.9482
#> 
#> (1,3,3,.,.) = 
#>   6.9329  6.3405
#>   3.6646 -13.5643
#>  -2.9506  5.9330
#>  -2.8361 -7.1243
#>   0.7209 -3.3815
#> 
#> (2,3,3,.,.) = 
#>   1.5262  2.7090
#>   5.2327  2.3287
#>   6.2155 -2.0571
#>  -4.2310 -4.1960
#>   3.1807  2.8638
#> 
#> (3,3,3,.,.) = 
#>   0.5015  2.4013
#>   9.8059 -7.5275
#>  -0.3134  3.5004
#>   2.5308  1.6895
#>   7.3772 -7.0387
#> 
#> (1,1,4,.,.) = 
#>  -13.6532   4.4051
#>  -10.4147   5.0181
#>   -1.1286   1.7529
#>    2.8230  -5.6151
#>   -7.9838  -2.6413
#> 
#> (2,1,4,.,.) = 
#>    4.0084   7.7560
#>   -0.2307  -0.0883
#>    4.3662  10.6252
#>   -2.3291  -0.1440
#>    1.5278   0.1987
#> 
#> (3,1,4,.,.) = 
#>  -9.1532 -6.4304
#>   6.9663 -1.3172
#>  -3.9041 -5.8065
#>   0.3933 -2.6468
#>   8.6516  4.8018
#> 
#> (1,2,4,.,.) = 
#>  -0.3304  5.5103
#>  -10.1824  6.2636
#>   2.9708 -8.4041
#>  -10.9114 -7.5795
#>  -2.4482 -5.0504
#> 
#> (2,2,4,.,.) = 
#>  -8.0483  0.7491
#>  -5.7506 -2.1713
#>   2.0948  6.2371
#>  -4.2344 -0.3634
#>  -5.2435 -3.3073
#> 
#> (3,2,4,.,.) = 
#>   4.8102  5.9549
#>   0.1113  4.5301
#>   7.1013  5.4995
#>  -4.5926 -3.2111
#>   6.9338  6.1466
#> 
#> (1,3,4,.,.) = 
#>    6.1493   4.0264
#>    1.4464  13.3106
#>    1.6339   6.9375
#>    2.8700   6.0982
#>   -1.5679   1.1448
#> 
#> (2,3,4,.,.) = 
#>   2.6702  3.4808
#>   6.4623  4.0797
#>   1.3476  4.3666
#>  -3.1097 -6.0254
#>   0.4799 -3.6648
#> 
#> (3,3,4,.,.) = 
#>   3.7527  5.6061
#>  -2.6108  2.9753
#>  -4.2929 -6.6872
#>   3.3525 -9.4171
#>  -5.5593 -3.8154
#> 
#> (1,1,5,.,.) = 
#>  -6.2914 -3.7980
#>  -0.1560  2.0582
#>  -1.9810  0.4032
#>   2.8557  3.1879
#>  -7.8761  5.1946
#> 
#> (2,1,5,.,.) = 
#>  -1.4126  1.8956
#>   6.4759 -3.8589
#>   7.1188 -1.0978
#>  -5.6880  1.9322
#>  -6.6137 -5.2791
#> 
#> (3,1,5,.,.) = 
#>   0.3488 -3.0268
#>   2.4352 -3.3522
#>   9.2019 -6.0556
#>  -2.2342 -1.8149
#>   2.2566 -1.4601
#> 
#> (1,2,5,.,.) = 
#>  -2.9594 -1.2969
#>   2.0627 -1.3316
#>   3.2959  2.1453
#>   4.5162 -6.4966
#>  -7.8497  5.9245
#> 
#> (2,2,5,.,.) = 
#>  -6.4284  7.8053
#>   1.0767 -1.3863
#>   8.8502  8.3183
#>   0.7857  1.9681
#>  -3.2123 -1.1060
#> 
#> (3,2,5,.,.) = 
#>  -3.9549 -5.9318
#>   1.7523  2.4051
#>   3.3194  5.8822
#>   0.0958  5.6401
#>  -0.5973  1.8942
#> 
#> (1,3,5,.,.) = 
#>   0.6352  1.7775
#>  -1.8487 -0.7606
#>   4.7322  8.5498
#>  -0.2256  0.7480
#>  -8.2643 -1.3865
#> 
#> (2,3,5,.,.) = 
#>   1.4332  0.5647
#>   3.8933  7.6698
#>  -9.0813  0.8803
#>  -0.8624  1.6519
#>   3.8879  2.2301
#> 
#> (3,3,5,.,.) = 
#>  -6.1772 -0.0762
#>   4.7082  2.9675
#>   3.5757 -5.9739
#>  -0.3279  1.5277
#>  -4.3354 -0.5859
#> [ CPUFloatType{3,3,5,5,2} ]