Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   -3.5686  -0.7575
#>   -0.9110   3.0255
#>   11.9251   0.8822
#>   11.8014   5.3858
#>    5.8462 -12.1556
#> 
#> (2,1,1,.,.) = 
#>   4.7089 -3.9589
#>  -8.4639 -3.9255
#>  -0.4952 -11.0552
#>  -0.1713 -5.6661
#>  -5.7493 -1.4321
#> 
#> (3,1,1,.,.) = 
#>  -7.2831  0.6017
#>   0.8504 -3.4139
#>   0.0712  7.0966
#>   3.9787 -0.9188
#>   7.0125 -6.0596
#> 
#> (1,2,1,.,.) = 
#>   1.7555  1.8364
#>   3.2766  1.2616
#>   2.3623  8.5845
#>   3.9022  7.0934
#>   2.8848  1.5320
#> 
#> (2,2,1,.,.) = 
#>   6.4026  9.1591
#>  -0.3073 -5.6671
#>   3.6115 -0.4895
#>  -10.6002 -2.3799
#>  -8.2712 -6.7868
#> 
#> (3,2,1,.,.) = 
#>  -5.7464  6.6129
#>  -0.5858  1.9289
#>   9.6124  5.2303
#>   2.7807 -2.0940
#>   2.7963 -4.3982
#> 
#> (1,3,1,.,.) = 
#>   1.3732 -4.0567
#>   1.6165 -8.0190
#>  -2.8664  3.5672
#>   5.9506 -0.5644
#>  -7.0809  1.9352
#> 
#> (2,3,1,.,.) = 
#>   8.8405  5.2406
#>  -1.4716 -5.5854
#>   8.5659  0.9253
#>   3.2954 -1.5305
#>   3.8698 -10.1214
#> 
#> (3,3,1,.,.) = 
#>  -2.9906  3.1035
#>  -2.4267  0.8469
#>   0.4349 -2.6360
#>   5.0007  1.6405
#>  -8.2372 -2.8690
#> 
#> (1,1,2,.,.) = 
#>   1.6505 -2.8279
#>  -3.1168 -5.8456
#>  -4.9219 -3.6338
#>   0.7485  4.2856
#>   0.3340 -1.0839
#> 
#> (2,1,2,.,.) = 
#>  -4.5411 -1.3995
#>   0.6041  0.6591
#>   2.6971  0.0667
#>   0.1509 -3.2606
#>  -1.6997 -5.7515
#> 
#> (3,1,2,.,.) = 
#>  -0.9648  2.3399
#>   3.2859  2.6885
#>   3.5364 -9.5266
#>  -6.3387  4.7638
#>  -0.9046  2.6541
#> 
#> (1,2,2,.,.) = 
#>   0.8476  0.1293
#>   2.4993 -11.7474
#>  -8.2536 -1.7945
#>   1.4565 -1.7748
#>   1.6302 -5.9520
#> 
#> (2,2,2,.,.) = 
#>   8.4007 -0.6091
#>  -7.3150  6.7983
#>  -2.7567  5.7257
#>   0.0162 -4.8731
#>   5.6091 -0.1334
#> 
#> (3,2,2,.,.) = 
#>   3.8595 -6.5510
#>  -7.0130  1.7539
#>   5.8311 -5.2801
#>  -0.3624  3.3297
#>  -1.0575  8.0131
#> 
#> (1,3,2,.,.) = 
#>   2.8984 -1.4383
#>  -2.3224  0.3968
#>  -4.7943  9.7074
#>  -0.7067 -0.1323
#>   4.0373 -4.3214
#> 
#> (2,3,2,.,.) = 
#>   4.1119 -5.5099
#>   5.7705  0.1910
#>  -1.1444  1.3298
#>   0.5945  1.7006
#>  -7.2836  2.9919
#> 
#> (3,3,2,.,.) = 
#>   -3.8975   4.4903
#>   -6.4195   4.3905
#>   -7.2374  11.6386
#>   -2.9600  -0.9243
#>    1.8684   2.2974
#> 
#> (1,1,3,.,.) = 
#>   4.5520 -4.3530
#>  -5.8284 -2.1102
#>  -3.6668 -7.7019
#>   8.1796 -19.0700
#>   9.5884  4.8352
#> 
#> (2,1,3,.,.) = 
#>   13.9761   3.4310
#>    0.4168  -6.5841
#>   -0.8708   7.1304
#>   -6.1184   3.3108
#>   10.5837  -8.0658
#> 
#> (3,1,3,.,.) = 
#>   7.0383  4.9793
#>   6.1311  0.5772
#>   0.5536 -10.8963
#>  -8.7085  4.4828
#>   2.9431 -2.5743
#> 
#> (1,2,3,.,.) = 
#>    4.1403   1.8223
#>    0.7537   0.8800
#>   -3.9671  -2.6145
#>   14.4936  -0.9029
#>    1.6383  10.4394
#> 
#> (2,2,3,.,.) = 
#>   4.4319  0.7040
#>   0.3135  5.0567
#>   1.3790  2.3072
#>  -1.2037 -2.0628
#>  -0.8419  3.1004
#> 
#> (3,2,3,.,.) = 
#>   -5.0752  -7.0265
#>    2.7095  11.5003
#>    4.8393  -5.2051
#>    1.2874  -6.7991
#>   -0.0882   1.2251
#> 
#> (1,3,3,.,.) = 
#>   5.1311 -0.5998
#>  -0.1163  0.8143
#>  -5.7811  0.3771
#>   7.0823 -1.5296
#>  -6.8239  3.0566
#> 
#> (2,3,3,.,.) = 
#>  -3.9121  4.2962
#>   7.0786  0.5725
#>   3.3395 -4.7522
#>  -1.2299  3.4169
#>  -2.0444 -7.8404
#> 
#> (3,3,3,.,.) = 
#>   0.3150 -0.1373
#>   4.7729  3.5093
#>   4.5809  0.5698
#>  -0.8044 -12.8542
#>   0.8985  7.5519
#> 
#> (1,1,4,.,.) = 
#>  -4.0290  1.0845
#>  -0.0392 -5.5119
#>   7.0016 -1.1269
#>  -11.4933  0.8587
#>   0.2535 -3.7267
#> 
#> (2,1,4,.,.) = 
#>    3.0707  -0.5112
#>    3.6116  10.5010
#>   -6.0431   6.8904
#>    6.3959   0.7005
#>   -5.3762  -4.8846
#> 
#> (3,1,4,.,.) = 
#>   1.0899 -3.3002
#>  -1.8715 -5.4116
#>   1.1477  1.5257
#>   3.4016 -5.9017
#>  -7.3450 -2.7497
#> 
#> (1,2,4,.,.) = 
#>   -4.4946  -0.0431
#>   -1.6383  -4.3185
#>    2.9340  -5.2805
#>   -9.9854   5.2759
#>   11.8919  -3.9943
#> 
#> (2,2,4,.,.) = 
#>   -4.7635   1.2913
#>    5.9620  -5.5513
#>   -8.0413  -8.4340
#>   11.9817   1.7833
#>   -1.7789  -1.3288
#> 
#> (3,2,4,.,.) = 
#>  -4.7120 -6.4397
#>   7.8180 -0.4252
#>   1.5676  3.6812
#>   7.2914 -0.5529
#>   3.5581 -2.2597
#> 
#> (1,3,4,.,.) = 
#>   8.5806  0.0611
#>  -0.4777  7.7718
#>  -4.0794 -6.1907
#>  -1.4671  5.2149
#>   1.0252 -8.5149
#> 
#> (2,3,4,.,.) = 
#>   -9.8472   5.6967
#>   -1.6947   9.5993
#>   -1.4956   4.0334
#>   -1.3865  11.1374
#>    9.8793   4.2991
#> 
#> (3,3,4,.,.) = 
#>   1.8941 -3.0498
#>  -0.5273  5.9420
#>  -8.1000  0.7917
#>  -4.0924  6.8378
#>  -3.5923 -4.5370
#> 
#> (1,1,5,.,.) = 
#>  -1.7056  9.2700
#>   2.1274  1.8812
#>   4.0498 -11.0172
#>   2.9746 -0.4427
#>   7.6162  9.7190
#> 
#> (2,1,5,.,.) = 
#>  -1.6148  9.8823
#>  -2.2223 -5.7224
#>  -9.0196 -1.7276
#>   4.7798  0.5862
#>   1.3281  0.8029
#> 
#> (3,1,5,.,.) = 
#>   -5.5273   7.8603
#>   -0.1868  -1.6501
#>   -2.8454   3.7523
#>   10.0059  -4.9990
#>    4.3484  -6.8718
#> 
#> (1,2,5,.,.) = 
#>  -2.2279 -3.9872
#>  -3.4575 -4.0907
#>   3.4209 -8.6413
#>   5.5585  3.9019
#>   1.2711 -2.7452
#> 
#> (2,2,5,.,.) = 
#>   1.7661 -6.5353
#>  -0.8081 -6.9792
#>  -5.1229  5.2656
#>   7.4427  5.2206
#>   4.4274  7.1338
#> 
#> (3,2,5,.,.) = 
#>   -9.5314   1.8814
#>   -1.1682  -6.7481
#>    6.2892  -5.7666
#>   -1.1676   3.0302
#>    8.6887  16.3532
#> 
#> (1,3,5,.,.) = 
#>  -5.6137  2.0296
#>   4.6979 -3.9769
#>  -0.7540 -5.3442
#>  -3.3033  1.3738
#>   3.3125 -4.9288
#> 
#> (2,3,5,.,.) = 
#>   7.8575 -5.3907
#>   4.6923 -0.7001
#>  -0.7059  2.9104
#>  -6.5708  5.6068
#>  -3.4356  5.0028
#> 
#> (3,3,5,.,.) = 
#>  -9.7962 -4.0961
#>   2.9536 -3.6162
#>  -1.1836  0.3741
#>  -3.6544 -5.1600
#>  -1.5882  2.3636
#> [ CPUFloatType{3,3,5,5,2} ]