Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -1.1481 -4.3297
#>  -4.7188 -6.2804
#>  -1.3097 -4.8588
#>  -0.6476 -7.7830
#>   5.8857  0.0997
#> 
#> (2,1,1,.,.) = 
#>  -1.6040  2.3841
#>   2.0651 -0.7709
#>   3.9230  8.8587
#>   5.2396 -3.9834
#>   2.9333 -2.8714
#> 
#> (3,1,1,.,.) = 
#>   0.3662  1.1582
#>   3.3350 -1.2566
#>  -6.8803  1.0332
#>   0.6336 -4.3311
#>  -8.1302 -1.4290
#> 
#> (1,2,1,.,.) = 
#>   9.9370 -4.7862
#>   1.7947  8.3546
#>   7.8480 -4.3197
#>  -4.5602 -4.6462
#>   9.8942 -1.0242
#> 
#> (2,2,1,.,.) = 
#>   5.4458 -1.6464
#>   0.9909 -0.2878
#>  -3.9934  7.8341
#>   0.6661  0.3599
#>   9.6997 -2.0700
#> 
#> (3,2,1,.,.) = 
#>   -0.2640  13.3921
#>   -3.9612   1.9535
#>   -2.7264  -6.7090
#>    6.3232  -2.0279
#>    3.9858  -3.9658
#> 
#> (1,3,1,.,.) = 
#>  -4.8588 -0.8276
#>   4.9844 -1.4986
#>  -2.8926 -1.1695
#>   5.5408 -3.3529
#>  -1.9728 -1.8652
#> 
#> (2,3,1,.,.) = 
#>   6.2058 -10.3060
#>   2.3539  0.2381
#>   2.5999  2.7840
#>   4.4128 -0.1960
#>  -1.7241 -1.4131
#> 
#> (3,3,1,.,.) = 
#>   3.4567 -0.4836
#>   6.1658  2.3352
#>   0.0856  0.4609
#>   3.7644 -5.1061
#>   1.0802 -4.8710
#> 
#> (1,1,2,.,.) = 
#>  -3.5388  6.0588
#>  -3.2331 -3.7648
#>  -3.3814 -2.3884
#>   1.9866  0.6786
#>   3.8671  0.4380
#> 
#> (2,1,2,.,.) = 
#>  -2.7273 -7.3688
#>   3.1946  1.7876
#>   5.9637 -6.5781
#>  -10.5504 -7.0482
#>   9.0913 -10.8848
#> 
#> (3,1,2,.,.) = 
#>   1.7073  4.0997
#>  -1.8936 -5.7653
#>  -2.6932 -3.5705
#>   2.1578  6.4609
#>   0.3697  4.2655
#> 
#> (1,2,2,.,.) = 
#>    3.2221  -3.5457
#>   -5.6733  -1.1672
#>   -6.0391  14.1264
#>   -1.8346  -4.6253
#>    1.4615   1.0004
#> 
#> (2,2,2,.,.) = 
#>   7.8673  2.4643
#>   6.2591  2.9252
#>  -0.3886 -0.8997
#>   0.1365 -4.4832
#>   3.5114  0.2806
#> 
#> (3,2,2,.,.) = 
#>    2.5126  -0.0145
#>    8.0643  -0.4143
#>    6.9933   0.4218
#>    7.4069   3.1231
#>   10.0771   0.4146
#> 
#> (1,3,2,.,.) = 
#>   8.7499  6.0461
#>  -6.3412 -2.5804
#>  -1.8398 -11.6373
#>   0.7424 -0.4836
#>  -3.3934  0.5697
#> 
#> (2,3,2,.,.) = 
#>  -0.8282  2.5507
#>   2.5071 -1.9104
#>  -6.8926 -1.4344
#>  -2.7887  0.8809
#>   1.3705 -3.8546
#> 
#> (3,3,2,.,.) = 
#>  -1.3417 -2.2560
#>  -0.9487 -2.3724
#>  -8.4024  6.2669
#>  -9.3354 -14.5017
#>   3.7803 -3.7519
#> 
#> (1,1,3,.,.) = 
#>  -1.9732 -8.0447
#>  -4.0135  4.3342
#>  -8.5510 -3.9217
#>   1.7348  5.0812
#>  -1.3394  5.5100
#> 
#> (2,1,3,.,.) = 
#>  -2.3784 -3.2535
#>  -0.5325 -1.7479
#>   0.5439  1.1196
#>  -1.7045 -6.2396
#>   1.8775 -5.5534
#> 
#> (3,1,3,.,.) = 
#>  -7.0758 -4.6158
#>  -4.4565  3.7113
#>  -2.5085 -0.5332
#>   2.2014 -1.8736
#>  -6.4190  0.8344
#> 
#> (1,2,3,.,.) = 
#>  -1.6565  7.4865
#>   7.9111  8.2888
#>   1.7455 -7.7078
#>  -8.7524  3.5801
#>   3.2935  2.1932
#> 
#> (2,2,3,.,.) = 
#>  -0.4705 -8.0645
#>  -2.7541  5.7445
#>   3.0258 -0.3652
#>   4.6185 -9.1709
#>  -3.3521  4.3578
#> 
#> (3,2,3,.,.) = 
#>  -9.7571  5.1088
#>   0.5665  3.9230
#>   0.8925 -0.7799
#>  -2.2061  9.9910
#>   0.5800  0.5737
#> 
#> (1,3,3,.,.) = 
#>   -7.3047   4.8323
#>   -3.9106   3.1690
#>   -6.3493  -0.1799
#>    2.7435  13.2830
#>    3.4891  -3.3635
#> 
#> (2,3,3,.,.) = 
#>   5.1295 -1.2111
#>  -5.8640 -4.0844
#>  -0.3176  2.4835
#>  -6.5782 -4.5152
#>  -0.8305  0.7476
#> 
#> (3,3,3,.,.) = 
#>  -5.4378 -5.3350
#>  -2.3722 -11.4782
#>   7.3945  0.4418
#>   1.8523 -0.8628
#>   1.4826 -7.1815
#> 
#> (1,1,4,.,.) = 
#>   2.4687  5.1131
#>   6.3423 -4.0122
#>   4.2059  0.0545
#>  -3.0227 -8.9330
#>   7.3082  3.1296
#> 
#> (2,1,4,.,.) = 
#>  -1.9773  4.5083
#>  -2.9370  0.5696
#>  -1.3116  0.0959
#>   3.0593  2.6283
#>   0.7708 -1.4834
#> 
#> (3,1,4,.,.) = 
#>  -0.4809  8.7561
#>  -0.4841  3.7019
#>  -2.3270  0.4256
#>   3.7321 -10.0876
#>  -2.0148  6.4611
#> 
#> (1,2,4,.,.) = 
#>   -2.6632   5.7187
#>   11.6131   2.1020
#>    1.8945   2.9998
#>   -0.9931 -15.3454
#>    1.6627   1.9919
#> 
#> (2,2,4,.,.) = 
#>  -0.7808 -2.8565
#>   8.3442  0.4109
#>   2.7123  0.6158
#>   1.3526  5.0902
#>   3.2664 -1.5515
#> 
#> (3,2,4,.,.) = 
#>    5.0697   0.5581
#>  -11.7244   1.5131
#>   -4.9459   3.7161
#>   -3.6735   2.2271
#>   10.1182  -0.3320
#> 
#> (1,3,4,.,.) = 
#>   3.6115 -1.6901
#>  -1.0491  6.2340
#>   2.3156 -0.1393
#>  -0.2285  1.8952
#>  -0.8674 -1.8877
#> 
#> (2,3,4,.,.) = 
#>   5.8264  6.5763
#>  -5.1067  2.8225
#>   4.7281  5.3284
#>   3.8772 -8.7982
#>   7.9865  3.5250
#> 
#> (3,3,4,.,.) = 
#>    5.6288  -2.1604
#>    5.2852  12.1708
#>   -6.5885  -1.3408
#>   -1.0228  -3.8949
#>    4.9946   5.1332
#> 
#> (1,1,5,.,.) = 
#>  -3.6772 -2.6791
#>   9.3151  6.3421
#>   1.0636 -2.4348
#>   0.0648  0.2960
#>  -9.0852 -0.3164
#> 
#> (2,1,5,.,.) = 
#>  -6.7774  5.5203
#>   8.8951  0.4216
#>   1.4675 -0.9553
#>  -4.2931  2.7152
#>  -8.8658  0.0596
#> 
#> (3,1,5,.,.) = 
#>   -0.8913  -3.6577
#>   -2.9250   2.1499
#>    1.9352  -0.6881
#>   12.8338   2.7566
#>   -8.3175   0.9609
#> 
#> (1,2,5,.,.) = 
#>   -2.9347   0.3892
#>   -1.4719  -6.7884
#>    0.1891   8.8247
#>   11.0228   8.0462
#>   -3.6506   3.0815
#> 
#> (2,2,5,.,.) = 
#>    2.6453  -9.5708
#>   -1.4926   2.9505
#>    9.6809  -4.5462
#>   10.8256   6.3338
#>    4.1002  -1.5479
#> 
#> (3,2,5,.,.) = 
#>  -1.0330  5.5573
#>  -2.3059 -1.8269
#>  -3.2600  6.0368
#>   4.6868  2.6289
#>   4.2373  0.1010
#> 
#> (1,3,5,.,.) = 
#>   6.1920  5.8462
#>   0.6780  4.9170
#>   2.4530 -14.6991
#>   8.8984  0.1436
#>  -0.9811  3.2451
#> 
#> (2,3,5,.,.) = 
#>  -3.2560  0.5176
#>   8.5553 -0.6800
#>  -7.4328 -0.9606
#>  -6.5130 -0.3419
#>   2.8921 -2.0966
#> 
#> (3,3,5,.,.) = 
#>   4.2309  0.3529
#>   4.4683  8.9331
#>   2.9729 -1.8606
#>   1.4575  3.8359
#>  -5.0445 -14.9692
#> [ CPUFloatType{3,3,5,5,2} ]