Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -1.0384 -6.1587
#>  -1.0615 -1.4415
#>   9.2712 -5.2620
#>   1.4542 -7.9069
#>  -4.9976 -8.6061
#> 
#> (2,1,1,.,.) = 
#>  -6.0561  5.1335
#>  -1.1490 -3.9663
#>  -0.0058 -4.5472
#>  -2.7145 -8.6047
#>  -3.9613  3.2926
#> 
#> (3,1,1,.,.) = 
#>   0.2758 -2.5975
#>   0.2778 -2.6216
#>   3.9000 -2.3022
#>  -2.2731 -0.1212
#>  -0.2957 -5.7145
#> 
#> (1,2,1,.,.) = 
#>   4.5483  9.9056
#>  -2.8430 -2.8520
#>  -6.6753 -3.3746
#>  -3.6461  9.0963
#>  -0.0045  6.3670
#> 
#> (2,2,1,.,.) = 
#>    8.7227  -1.2409
#>    9.9848   0.7758
#>   -4.4883   3.8475
#>   -6.1900  -0.8637
#>   -5.1797  11.0081
#> 
#> (3,2,1,.,.) = 
#>  -2.6396 -10.2788
#>   3.7941 -2.2107
#>  -0.8698  1.4420
#>   5.9698  1.0226
#>   4.5721 -3.4027
#> 
#> (1,3,1,.,.) = 
#>  -1.1211 -7.6841
#>   0.8187  1.7869
#>  -1.3844 -7.2692
#>  -11.6808  5.7017
#>   6.0401  8.5203
#> 
#> (2,3,1,.,.) = 
#>   9.9043 -8.8612
#>   5.8446 -5.1748
#>   0.0114 -7.8828
#>   4.4882 -3.4659
#>  -3.0172  0.8709
#> 
#> (3,3,1,.,.) = 
#>  -1.3448 -1.1314
#>   1.7031  2.9445
#>   8.9323  1.1808
#>   5.7438 -0.2184
#>   2.4642  4.4519
#> 
#> (1,1,2,.,.) = 
#>   9.1516  7.1533
#>   4.9348  3.9512
#>  -1.2362  2.7790
#>   7.4923  1.6325
#>   1.8666  7.3507
#> 
#> (2,1,2,.,.) = 
#>   8.4702  7.4189
#>  -2.0988 -1.3479
#>  -2.9435  1.7058
#>  -9.8016 -0.0928
#>   0.1647  5.2544
#> 
#> (3,1,2,.,.) = 
#>   -6.8588  -4.0609
#>    6.5212  14.5668
#>    6.4299   0.1073
#>  -12.8726   3.9909
#>    4.6998  -3.4491
#> 
#> (1,2,2,.,.) = 
#>    0.1773  -0.8506
#>    2.0110  -0.3512
#>   -8.7091  11.7349
#>    4.7372 -10.4853
#>   -4.6923  -5.9272
#> 
#> (2,2,2,.,.) = 
#>   -0.6132  -3.7200
#>   12.3590  -2.8159
#>    6.4834   4.4668
#>    0.3464  -4.0384
#>   -2.0003   4.3548
#> 
#> (3,2,2,.,.) = 
#>   -2.8336   5.9111
#>    2.2688  10.0888
#>   -1.0320  -1.5736
#>   -6.1492   0.5599
#>    1.3804  -1.0394
#> 
#> (1,3,2,.,.) = 
#>   0.3353  0.3218
#>  -6.3596  2.1246
#>   6.8508  2.1810
#>   1.4644 -4.2533
#>  -4.1313 -8.1800
#> 
#> (2,3,2,.,.) = 
#>    2.4041   1.8391
#>   13.4807  11.4190
#>   -1.6303   0.6522
#>    5.1852  -4.8535
#>    4.5425   4.6395
#> 
#> (3,3,2,.,.) = 
#>    2.6775  10.3401
#>  -10.0568  10.0742
#>    4.3540  -2.1784
#>   -3.5257  -1.6395
#>   -1.7252  -0.4881
#> 
#> (1,1,3,.,.) = 
#>   11.5881   2.3851
#>    1.8668   5.7783
#>    5.3618   7.9721
#>   -0.5439  -1.3317
#>    0.9493   6.2107
#> 
#> (2,1,3,.,.) = 
#>   1.5744  5.5975
#>  -5.8934  1.4273
#>  -10.1965  4.9231
#>   6.0822 -5.4652
#>  -9.9620  5.4097
#> 
#> (3,1,3,.,.) = 
#>   4.7346  8.8494
#>   4.4375  0.9101
#>   4.8890 -7.4432
#>   2.1319 -0.5947
#>   2.8731 -2.1297
#> 
#> (1,2,3,.,.) = 
#>   5.1573 -7.0380
#>  -5.2382 -1.0651
#>  -2.4547 -1.0326
#>   4.5580  2.2170
#>  -1.0331  9.6623
#> 
#> (2,2,3,.,.) = 
#>  -5.0264 -1.1786
#>  -1.7166 -2.4167
#>   7.1171  1.5796
#>   5.2548 -3.4157
#>  -2.9674 -1.8949
#> 
#> (3,2,3,.,.) = 
#>   2.4943  7.4627
#>   9.3321  7.6037
#>  -2.7005 -1.8666
#>   2.9436 -3.2694
#>  -10.0785  3.5874
#> 
#> (1,3,3,.,.) = 
#>   7.6765 -18.3679
#>  -3.8349  6.2456
#>   1.8638  2.1820
#>   2.8326  2.4645
#>  -4.2196 -0.7073
#> 
#> (2,3,3,.,.) = 
#>    1.8817  -1.0717
#>    5.4726  -0.5722
#>   10.7869  -6.5820
#>   -2.0584   1.6875
#>   -0.5640   4.7224
#> 
#> (3,3,3,.,.) = 
#>  -9.9337  2.6101
#>  -3.6538 -1.8038
#>  -10.9493 -2.3127
#>   7.1517  0.0032
#>  -6.9123 -2.5171
#> 
#> (1,1,4,.,.) = 
#>  -0.3775  3.6209
#>   6.1700 -3.4734
#>   2.7476 -0.5118
#>   0.5693 -6.3525
#>   5.5211 -1.4945
#> 
#> (2,1,4,.,.) = 
#>   0.1507  3.2414
#>   0.7060 -0.4267
#>   1.9502  3.9942
#>  -12.6060  2.9621
#>  -0.8166 -0.6526
#> 
#> (3,1,4,.,.) = 
#>   6.4443 -0.9175
#>  -0.5056  5.4403
#>  -0.5498  0.0530
#>  -4.9294  5.9108
#>  -7.8495 -3.0845
#> 
#> (1,2,4,.,.) = 
#>   1.2117 -1.0891
#>   3.6326 -6.2690
#>   5.8069  2.4480
#>   0.1324 -8.3948
#>  -2.2521  8.8168
#> 
#> (2,2,4,.,.) = 
#>   3.0319  3.2715
#>   2.9437 -1.8247
#>   7.2712  8.1436
#>   2.9974  1.0464
#>   0.0196  4.5070
#> 
#> (3,2,4,.,.) = 
#>  -1.4379 -1.4972
#>  -0.3461 -1.8270
#>  -1.6612 -1.5845
#>   0.9988 -2.9751
#>  -2.6848  2.5213
#> 
#> (1,3,4,.,.) = 
#>  -0.3375 -6.9659
#>  -7.5405  5.8664
#>   2.1742  5.6089
#>   0.4461  4.1198
#>   4.3604 -2.4046
#> 
#> (2,3,4,.,.) = 
#>   5.8392  4.0256
#>  -0.1561 -6.4677
#>   0.5422  2.3773
#>   1.9313 -11.5171
#>   0.8714 -7.0798
#> 
#> (3,3,4,.,.) = 
#>   -3.0377  -0.8727
#>   10.1549  -5.7045
#>   10.5575  -3.3163
#>   -6.2932  -2.7636
#>   -0.9400   2.1597
#> 
#> (1,1,5,.,.) = 
#>   -2.5472   4.5632
#>   -4.7861  -5.2718
#>    0.8535   7.7773
#>   10.9148   3.5813
#>   -1.3347  -8.2269
#> 
#> (2,1,5,.,.) = 
#>   0.8658 -4.1016
#>   6.0096  6.9074
#>  -4.0570  5.9583
#>   4.2062 -5.5420
#>  -4.8079 -1.4389
#> 
#> (3,1,5,.,.) = 
#>  -9.1883 -4.0213
#>   1.8579 -6.0041
#>  -0.5521 -8.4692
#>   4.5119  1.1459
#>  -1.6726 -9.9645
#> 
#> (1,2,5,.,.) = 
#>   0.8202 -0.8812
#>  -3.7412  2.5875
#>  -2.2793 -5.1110
#>  -3.5529 -1.9516
#>  -2.1914  1.8079
#> 
#> (2,2,5,.,.) = 
#>   -7.3619   4.6347
#>   10.3466  -1.6812
#>    3.4903   3.9215
#>    2.4918  10.7266
#>   -1.2656   4.9389
#> 
#> (3,2,5,.,.) = 
#>    4.5740  -9.7328
#>   -9.2921   4.4611
#>    7.4413   3.0222
#>   -2.7911  10.7842
#>   -6.2412   2.0884
#> 
#> (1,3,5,.,.) = 
#>  -5.4536  4.8640
#>  -0.1327 -3.9056
#>  -1.4054  0.1277
#>   0.0533 -9.6020
#>  -4.6667 -6.7094
#> 
#> (2,3,5,.,.) = 
#>  -4.2498 -6.8149
#>  -1.2650  1.0567
#>  -7.1767 -2.6874
#>  -2.4965  3.0355
#>   0.3435  3.5718
#> 
#> (3,3,5,.,.) = 
#>  -9.4431 -1.9084
#>   0.5088  7.2661
#>   4.3385 -3.4697
#>  -0.8239 -1.5861
#>  -5.5470  0.6133
#> [ CPUFloatType{3,3,5,5,2} ]