Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   0.6538  2.3902
#>  -0.0407 -2.1009
#>  -1.7075  0.2743
#> 
#> (2,.,.) = 
#>  -7.3254 -4.1155
#>  -3.4979  4.6414
#>  -1.9202  6.4543
#> 
#> (3,.,.) = 
#>   0.7929  2.8143
#>  -1.5744 -1.1399
#>   1.8256  0.9958
#> 
#> (4,.,.) = 
#>  -1.1778  4.2228
#>  -1.2220 -6.5910
#>  -4.3523 -0.8053
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -1.7641  1.3280
#>  -1.5838 -1.2976
#>  -1.5386  1.7298
#> 
#> (2,.,.) = 
#>   2.0498 -1.6429
#>  -2.4247 -0.8092
#>  -2.6982  0.4276
#> 
#> (3,.,.) = 
#>   2.4875  1.2743
#>   0.7762 -0.3228
#>   1.5976 -1.0947
#> 
#> (4,.,.) = 
#>  -2.1193  1.4309
#>   3.1915  0.3287
#>   0.9316 -0.7884
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -5.9427  2.3830
#>  -0.9876  3.9555
#>  -7.3313  3.1490
#>   4.2801 -8.8746
#>  -7.3879  5.3391
#> 
#> (2,1,1,.,.) = 
#>    4.5545   3.1398
#>    6.6397  12.7704
#>  -11.8139  -3.3971
#>   -6.0994   3.0087
#>    0.8275   1.3478
#> 
#> (3,1,1,.,.) = 
#>   -0.1075   2.4478
#>   -6.8575   1.1242
#>   -6.0275   3.7266
#>    4.3421  11.8535
#>   -1.6037   7.5695
#> 
#> (1,2,1,.,.) = 
#>   2.8736 -2.1080
#>   1.0969 -7.3034
#>   9.1364 -0.4367
#>   2.2691 -3.3444
#>  -4.3820 -4.2208
#> 
#> (2,2,1,.,.) = 
#>  -0.9118  4.7446
#>   1.1206  7.9196
#>   1.6641  0.7576
#>   3.0925  1.1970
#>   8.5249 -1.4439
#> 
#> (3,2,1,.,.) = 
#>   4.8518  0.1233
#>  -6.2846  7.4662
#>  -1.7859  6.2408
#>   1.5846 -0.8276
#>   1.2933 -3.6956
#> 
#> (1,3,1,.,.) = 
#>   -2.0138  -1.6578
#>   -4.5817  -4.3853
#>   -4.8263  -1.1006
#>    1.6396  -3.9959
#>    0.3164  18.2475
#> 
#> (2,3,1,.,.) = 
#>    7.8291   2.7374
#>   -5.7064  -9.9621
#>   -2.8155  -0.1657
#>   -3.9620  12.4611
#>    3.4435   3.2326
#> 
#> (3,3,1,.,.) = 
#>   4.4183  3.6809
#>   3.2348 -0.1091
#>   6.8238 -4.1220
#>  -4.5586 -7.4389
#>   5.1245 -2.3675
#> 
#> (1,1,2,.,.) = 
#>   9.7920  0.6806
#>  -0.3190 -3.7368
#>   0.7023 -0.9383
#>   4.7008 -1.5883
#>  -3.9566  3.6726
#> 
#> (2,1,2,.,.) = 
#>  -3.0512  1.9221
#>  -0.2157  0.0452
#>  -2.9183 -4.3607
#>   8.7492  7.3246
#>   1.5165  9.6445
#> 
#> (3,1,2,.,.) = 
#>   -0.0073  -0.8282
#>   -0.6035  -5.8420
#>   12.4597  -2.4005
#>    2.2515  11.3698
#>    4.6597   4.6143
#> 
#> (1,2,2,.,.) = 
#>  -1.2314 -3.1506
#>   1.3068  1.5538
#>   8.0836 -3.6069
#>   0.0689  7.7218
#>   1.6400  7.3772
#> 
#> (2,2,2,.,.) = 
#>   7.2325 -3.6465
#>   6.0541 -4.9124
#>  -2.4519  2.3710
#>  -9.0196  8.2937
#>  -6.5627 -7.3668
#> 
#> (3,2,2,.,.) = 
#>   4.8670  9.0995
#>   0.0583  4.9745
#>   3.3052 -2.3065
#>  -1.0784 -3.5113
#>  -1.3564 -2.9838
#> 
#> (1,3,2,.,.) = 
#>   0.1958 -0.7882
#>  -2.9442 -4.7447
#>  -0.6995 -5.9500
#>  -4.5564 -0.7496
#>   5.6657 -3.0886
#> 
#> (2,3,2,.,.) = 
#>   7.0723 -4.2332
#>   7.6992 -0.9180
#>   2.7281 -8.0427
#>   5.9747  1.6752
#>  -2.7002 -0.1596
#> 
#> (3,3,2,.,.) = 
#>   5.7039  1.1480
#>   2.7226  3.4380
#>  -4.1206  0.7478
#>  -1.0942 -2.8951
#>  -1.4779  1.0268
#> 
#> (1,1,3,.,.) = 
#>   7.1011 -0.2380
#>  -3.5223  3.5401
#>   5.7742  1.4789
#>  -4.6115  2.6932
#>  -2.7191 -4.2168
#> 
#> (2,1,3,.,.) = 
#>  -5.7986  6.4509
#>  -5.6827 -6.2617
#>  -4.3300  2.0698
#>   6.8867 -2.4145
#>   9.1154  3.2265
#> 
#> (3,1,3,.,.) = 
#>   2.4795  1.2525
#>   5.3633 -4.1375
#>  -2.7429 -0.0403
#>   1.9457  1.0570
#>   0.7156 -1.8286
#> 
#> (1,2,3,.,.) = 
#>  -0.0121 -3.2482
#>  -2.3066  3.7388
#>  -8.0017  3.0797
#>   6.7179  4.0203
#>  -2.2814 -5.3970
#> 
#> (2,2,3,.,.) = 
#>   0.6086 -5.4349
#>   3.6174  2.9092
#>  -7.8689 -13.5817
#>  -6.2091 -1.2536
#>  -0.3103  7.7553
#> 
#> (3,2,3,.,.) = 
#>  -10.8803   3.3297
#>   -6.5093  -1.5097
#>   -8.7847  -4.2053
#>  -11.8536  -7.2926
#>   -7.0811  -8.0697
#> 
#> (1,3,3,.,.) = 
#>  -2.9286  6.8577
#>   3.9069  0.3475
#>   0.2377 -3.9633
#>   7.2230  9.0634
#>   6.9915  5.3864
#> 
#> (2,3,3,.,.) = 
#>  -0.5471  3.3024
#>   5.6635  0.4806
#>   9.6620  8.0607
#>  -1.2398  1.4398
#>   1.1879  4.9579
#> 
#> (3,3,3,.,.) = 
#>    7.0490   9.4731
#>   -2.1840   7.1841
#>    1.6218   0.4988
#>   12.1390  10.0518
#>  -10.0381   3.5385
#> 
#> (1,1,4,.,.) = 
#>  -9.0473 -4.1510
#>   4.4224 -11.9244
#>   0.2593  5.7600
#>   5.8278  2.6686
#>  -3.3019 -3.9117
#> 
#> (2,1,4,.,.) = 
#>   0.3849  3.7750
#>   4.9458 -3.1538
#>   7.8825 -1.8965
#>   0.3016 -0.8280
#>  -1.3189 -2.3694
#> 
#> (3,1,4,.,.) = 
#>  -2.7014 -5.1757
#>   3.8298 -5.0397
#>  -2.6360 -0.4463
#>  -4.8332  3.6021
#>   3.3230 -1.7663
#> 
#> (1,2,4,.,.) = 
#>   8.1613 -2.2573
#>  -3.2551  1.3833
#>   4.3634  6.3142
#>   0.5248 -7.8280
#>   0.8889 -3.6975
#> 
#> (2,2,4,.,.) = 
#>   2.2364  3.2553
#>   3.8263  3.7596
#>  -3.3828  1.5186
#>   0.7146 -5.7372
#>  -3.8133 -0.6495
#> 
#> (3,2,4,.,.) = 
#>  -10.9589   5.8050
#>   -5.3726   5.7243
#>    4.5528  -6.0136
#>    7.4999  -0.2582
#>    3.3877   6.1032
#> 
#> (1,3,4,.,.) = 
#>   0.7270  6.1670
#>  -6.5006 -5.9085
#>   2.4775  1.7599
#>  -3.6453  4.0545
#>   5.4839  1.2570
#> 
#> (2,3,4,.,.) = 
#>  -6.8501 -5.9425
#>   1.7361  1.2303
#>  -0.0090  2.3497
#>  -1.8686 -0.6607
#>   2.4513 -1.8840
#> 
#> (3,3,4,.,.) = 
#>   5.9438 -1.2280
#>  -8.8390  9.0097
#>   2.8616 -1.3395
#>  -9.2470  0.0511
#>  -0.1519 -1.3970
#> 
#> (1,1,5,.,.) = 
#>  -3.3137  1.7789
#>   2.6371  2.7102
#>   4.4069 -4.4288
#>  -3.4736  1.9758
#>  -4.3432 -4.9600
#> 
#> (2,1,5,.,.) = 
#>  -2.1086 -0.6690
#>  -3.5679 -1.3710
#>  -3.9918 -4.4313
#>   3.3377 -1.4201
#>   4.3771  2.7254
#> 
#> (3,1,5,.,.) = 
#>  -9.5478  4.2810
#>  -5.3256 -5.4435
#>   1.0563 -2.9631
#>   3.6469 -2.4466
#>   2.9408  6.8321
#> 
#> (1,2,5,.,.) = 
#>  -3.1676  4.9672
#>   5.1714  2.4288
#>   0.3678  5.9836
#>  -1.7141  3.7235
#>  -5.0729 -2.1338
#> 
#> (2,2,5,.,.) = 
#>   3.2250  3.2594
#>   1.1617 -3.0566
#>  -1.8365 -0.7719
#>  -0.5639 -2.4706
#>   5.6257  9.9585
#> 
#> (3,2,5,.,.) = 
#>  -0.1548 -9.0428
#>   0.6164 -2.1338
#>  -2.3196  1.0116
#>   6.2667 -8.7648
#>  -6.9029 -2.2244
#> 
#> (1,3,5,.,.) = 
#>   6.4957 -2.0517
#>   0.6094 -3.5123
#>  -4.1952 -8.5566
#>   4.5907 -4.7177
#>  -7.8233 -2.9789
#> 
#> (2,3,5,.,.) = 
#>   1.4311 -0.1023
#>  -1.5866  3.9124
#>   3.7206 -2.0941
#>  -13.1653  5.1716
#>  -0.5996  7.7850
#> 
#> (3,3,5,.,.) = 
#>   0.8066 -1.9938
#>  -10.3948  0.4750
#>  -6.6004  6.0039
#>   0.0750  6.6125
#>  -1.6083 -0.2164
#> [ CPUFloatType{3,3,5,5,2} ]