Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   0.7945 -0.9114
#>  -0.6249 -1.0123
#>  -4.8704 -4.3114
#> 
#> (2,.,.) = 
#>   4.0883  5.1600
#>  -2.9950 -3.2388
#>  -0.7378  3.3267
#> 
#> (3,.,.) = 
#>   8.3165  1.1656
#>   0.5671 -4.9606
#>  -3.6608 -8.2291
#> 
#> (4,.,.) = 
#>   0.2304  5.4769
#>   1.4956  4.3491
#>  -1.9885  3.0172
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>   3.3574  2.7228
#>  -0.3893 -1.2156
#>  -2.8144 -1.5491
#> 
#> (2,.,.) = 
#>  -1.8013  0.4452
#>   1.5990 -0.1356
#>  -0.3798  1.2921
#> 
#> (3,.,.) = 
#>   1.1981 -2.5957
#>   0.3604 -1.7708
#>  -1.4512 -4.7211
#> 
#> (4,.,.) = 
#>  -1.9597 -1.4837
#>  -2.1950  2.1097
#>  -0.2250  0.6667
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>    1.2143   2.1742
#>    1.4577  -3.1231
#>    6.1894  11.9237
#>    5.7552  -6.5422
#>   -8.6975   1.8146
#> 
#> (2,1,1,.,.) = 
#>  -2.0848  3.6753
#>  -11.6592 -6.1599
#>   4.9849  1.2367
#>  -1.7702  4.2585
#>  -1.5247 -0.5281
#> 
#> (3,1,1,.,.) = 
#>  -7.2798  7.6163
#>  -10.9426 -6.3452
#>   4.9902 -0.1248
#>  -4.8032 -3.9598
#>  -5.1608 -0.1772
#> 
#> (1,2,1,.,.) = 
#>   -0.4154  -4.3870
#>   10.9117   1.7175
#>   -2.3082  -2.7608
#>    4.4576   2.7410
#>    6.2523  -7.6362
#> 
#> (2,2,1,.,.) = 
#>  -14.4488  -0.6255
#>    2.1863   2.8661
#>   -1.3967   5.5959
#>   -1.0677   3.0352
#>    3.1126  -3.1326
#> 
#> (3,2,1,.,.) = 
#>    0.5309  10.2751
#>   -2.9907   0.1418
#>    5.8892  -7.8431
#>   -5.3030   2.1175
#>   -6.2393   0.9570
#> 
#> (1,3,1,.,.) = 
#>  -2.5755  4.5326
#>   3.0747  2.3981
#>   1.4087  5.1213
#>   5.0870  5.7416
#>   1.0724 -0.1614
#> 
#> (2,3,1,.,.) = 
#>  -0.4362 -5.6240
#>   5.6630 -2.1170
#>   3.1697 -7.5868
#>  -3.7778 -14.9457
#>  -13.1525  6.2970
#> 
#> (3,3,1,.,.) = 
#>  -2.5137  1.9720
#>  -1.3676 -1.8902
#>   6.3205  3.1052
#>  -5.6248  6.5343
#>   3.6491  3.7806
#> 
#> (1,1,2,.,.) = 
#>  -0.8366  8.0817
#>   7.4917 -5.4754
#>   8.7307  6.9633
#>  -2.9353 -12.8533
#>   5.8858  7.8187
#> 
#> (2,1,2,.,.) = 
#>   3.6032 -1.9438
#>   7.2482  0.1710
#>   3.2319 -4.0271
#>  -0.2619  0.6414
#>   6.8515  4.4632
#> 
#> (3,1,2,.,.) = 
#>  -4.0053  1.3621
#>  -2.5134 -4.7299
#>   3.3001 -6.0116
#>  -1.2823  1.4673
#>  -8.6707  4.5058
#> 
#> (1,2,2,.,.) = 
#>    3.7314  -4.2835
#>   -1.8507  -3.1948
#>    3.8641   8.1611
#>    1.0452  10.3944
#>    3.7631   4.6116
#> 
#> (2,2,2,.,.) = 
#>  -3.8745  5.1776
#>  -3.9018 -8.7645
#>   0.5915 -2.8673
#>   0.1606  7.6933
#>  -9.8770 -6.6636
#> 
#> (3,2,2,.,.) = 
#>   1.5044  8.2641
#>  -1.6669 -6.0051
#>   4.5584 -3.3662
#>  -5.3077 -6.5522
#>  -2.7065 -3.9749
#> 
#> (1,3,2,.,.) = 
#>   -3.8999  10.5201
#>   -2.2787   5.6589
#>    0.7651  11.6901
#>   -3.3264  -7.3006
#>   -4.2228  -8.0243
#> 
#> (2,3,2,.,.) = 
#>   10.7280  -2.6968
#>    1.2852   0.2285
#>    3.9350   4.9735
#>    6.0806  -3.9743
#>   -0.0494  -4.3154
#> 
#> (3,3,2,.,.) = 
#>   -6.6224   3.3071
#>    4.0000  13.9375
#>    6.2501  -2.2826
#>   -0.4277   6.4399
#>    7.3649   1.4136
#> 
#> (1,1,3,.,.) = 
#>   10.4337   0.2586
#>   -2.1637  -1.6365
#>  -12.2609  -0.7159
#>   -0.2720   8.7454
#>    0.8956   1.9225
#> 
#> (2,1,3,.,.) = 
#>   9.0871  1.2287
#>  -6.6088  2.5213
#>   0.8227  3.3208
#>   3.4054 -2.5752
#>  -2.7543  3.9603
#> 
#> (3,1,3,.,.) = 
#>   7.0983 -0.6397
#>  -2.8697  0.3732
#>  -3.7263 -4.0337
#>  -7.5902 -0.4541
#>  -3.7807  0.6440
#> 
#> (1,2,3,.,.) = 
#>   1.1269  3.4771
#>  -9.9528  1.4445
#>   2.8782 -1.5219
#>   3.6589  1.1149
#>  -0.1971 -6.9450
#> 
#> (2,2,3,.,.) = 
#>    0.3106  -6.1158
#>   11.4082   6.5616
#>    0.0460  -7.2381
#>   -2.6575   0.8714
#>   -0.0096  -4.6866
#> 
#> (3,2,3,.,.) = 
#>  -0.2414 -3.5683
#>  -2.9758  3.0048
#>   0.2880  8.5061
#>  -0.3974 -4.3658
#>   1.5556  1.3638
#> 
#> (1,3,3,.,.) = 
#>   5.6348 -2.3079
#>  -3.9022 -4.0264
#>   2.4296 -1.0885
#>  -0.9375  3.0084
#>  -6.3297 -0.5131
#> 
#> (2,3,3,.,.) = 
#>   -0.7728  15.2383
#>    0.8735   2.2553
#>   -2.5199  -1.3575
#>   -6.2541   8.2985
#>    4.0309  -1.7885
#> 
#> (3,3,3,.,.) = 
#>   -4.9331  -1.0352
#>   -3.9118  -0.1667
#>    1.2149  -4.2679
#>    1.3245  13.6685
#>    7.7772   0.3306
#> 
#> (1,1,4,.,.) = 
#>   1.4920 -0.6440
#>  -1.6904 -0.8384
#>  -1.2818 -3.2467
#>  -1.2990  2.8279
#>   1.9830 -1.9103
#> 
#> (2,1,4,.,.) = 
#>   6.8750 -5.7230
#>  -6.8938  0.1096
#>   1.6133 -2.7832
#>   6.0887 -0.6115
#>   3.0232 -7.5660
#> 
#> (3,1,4,.,.) = 
#>  -6.0939 -10.8352
#>  -4.7440 -1.8385
#>   2.4017 -2.8250
#>  -2.1195 -1.0262
#>   1.4074  4.3985
#> 
#> (1,2,4,.,.) = 
#>  -9.4097 -2.3173
#>   4.0713 -0.7598
#>   1.1758  2.8020
#>   3.2354  8.3390
#>  -6.0176  6.1688
#> 
#> (2,2,4,.,.) = 
#>  -7.2796  2.7568
#>  -2.7606 -0.1358
#>  -5.1686  4.8345
#>   8.1415  0.0803
#>  -2.5770 -7.8998
#> 
#> (3,2,4,.,.) = 
#>  -1.4529 -0.7311
#>  -6.8433  1.4080
#>  -3.2933  9.5889
#>   5.9263  0.9474
#>  -5.1411 -2.9697
#> 
#> (1,3,4,.,.) = 
#>  -6.0790 -6.1804
#>  -10.2710 -0.0876
#>  -3.9623 -0.4611
#>  -6.7856  2.2815
#>  -7.4349  1.0840
#> 
#> (2,3,4,.,.) = 
#>   1.1367 -9.6872
#>  -4.8417 -1.2012
#>  -3.6490  3.7324
#>  -3.5171 -6.0426
#>  -1.7789  2.2112
#> 
#> (3,3,4,.,.) = 
#>   2.9251 -8.1866
#>  -0.7741  0.9303
#>   2.7847  2.2768
#>  -3.0724 -5.2096
#>   2.7685  5.1690
#> 
#> (1,1,5,.,.) = 
#>   -6.6914   5.2647
#>    1.1207   4.9504
#>    5.6371  10.8395
#>    0.7092  -1.6811
#>   -2.1857   0.3604
#> 
#> (2,1,5,.,.) = 
#>  -6.1799  2.1437
#>   7.2035  2.7068
#>  -0.7232 -2.6187
#>   1.3353 -9.1286
#>   7.0209 -4.9679
#> 
#> (3,1,5,.,.) = 
#>  -7.3302 -0.1717
#>   1.9238  0.0879
#>   1.9361 -8.1991
#>  -0.6378  0.3755
#>   4.7081 -8.7600
#> 
#> (1,2,5,.,.) = 
#>   4.6076  3.1750
#>  -4.0512  2.2308
#>  -2.9278  1.6574
#>   7.2857 -8.2523
#>  -6.3427 -3.6503
#> 
#> (2,2,5,.,.) = 
#>   5.7529  6.4722
#>   8.0358 -0.1027
#>   7.0215  0.8716
#>  -4.9017 -0.0774
#>   7.8037  4.1046
#> 
#> (3,2,5,.,.) = 
#>    6.2396   5.7240
#>   13.2659   2.7884
#>   -7.0918 -10.6518
#>   -3.2653   2.5056
#>   -4.5639   5.2769
#> 
#> (1,3,5,.,.) = 
#>   -6.1417  -1.0400
#>   -4.0190   3.4165
#>   11.5925   3.0827
#>   -5.8649  -0.6269
#>   -4.8381   0.8111
#> 
#> (2,3,5,.,.) = 
#>   2.0095 -3.2953
#>  -2.9784  7.7987
#>   3.7636 -0.1047
#>  -10.3683 -2.7364
#>   2.1098 -8.9063
#> 
#> (3,3,5,.,.) = 
#>  -3.8775 -8.9011
#>  -2.0926 -3.0090
#>   1.3002  0.6042
#>  -5.1572 -4.1419
#>   5.2865 -1.0919
#> [ CPUFloatType{3,3,5,5,2} ]