Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   3.8970  3.1487
#>   3.5089 -8.0428
#>  -4.2355  4.7304
#>   5.1422  8.4612
#>  -0.2829  6.5526
#> 
#> (2,1,1,.,.) = 
#>  -6.4914  0.7626
#>  -9.6798 -6.4868
#>   4.7413 -6.3207
#>   3.1433  6.0861
#>   1.4183  8.1368
#> 
#> (3,1,1,.,.) = 
#>   4.1796 -0.8876
#>  -2.7845 -0.2090
#>   6.4877 -2.0553
#>   1.6739 -1.4434
#>   6.7934  3.7603
#> 
#> (1,2,1,.,.) = 
#>   -8.0135  -3.4585
#>    3.2928   1.1881
#>    1.0519   7.3132
#>   -2.0995   0.7831
#>   11.4499   0.4932
#> 
#> (2,2,1,.,.) = 
#>   2.4901  2.9753
#>  -8.2115  4.1514
#>   6.0281  1.2984
#>  -7.5086  1.3346
#>   3.6055 -0.8337
#> 
#> (3,2,1,.,.) = 
#>    4.5424  -0.5391
#>   10.2617  -3.2754
#>   -1.8136   5.1917
#>   -4.6990  -0.9120
#>   -0.9396  -2.4928
#> 
#> (1,3,1,.,.) = 
#>  -4.4424 -3.1498
#>  -0.4156  0.1798
#>  -13.2501 -4.7913
#>  -0.4209  2.1745
#>  -0.4077 -2.1715
#> 
#> (2,3,1,.,.) = 
#>   1.8846  7.9342
#>   3.5554 -4.6065
#>   2.1314 -1.0630
#>  -1.8522  5.3271
#>  -4.2557 -6.4940
#> 
#> (3,3,1,.,.) = 
#>  -5.2358 -6.1865
#>  -13.5679  8.0294
#>   5.8208 -5.6350
#>   6.6610 -3.3221
#>  -2.0237 -1.4322
#> 
#> (1,1,2,.,.) = 
#>   3.3358  4.7716
#>   3.1996  3.2956
#>   2.6306  7.2663
#>  -2.2276 -8.0122
#>   2.0598 -2.9997
#> 
#> (2,1,2,.,.) = 
#>   0.5116  0.5725
#>   2.0175 -4.3590
#>  -1.4288  0.5337
#>   2.3421  0.0760
#>  -7.5872 -0.3140
#> 
#> (3,1,2,.,.) = 
#>   5.6348  1.3832
#>   6.6151  0.4225
#>   6.0133 -5.4770
#>  -7.6286 -6.4727
#>   3.6738  4.4512
#> 
#> (1,2,2,.,.) = 
#>   2.8345 -7.9496
#>  -1.2934  6.3027
#>  -2.5656  3.4925
#>  -4.2269  8.1202
#>   7.0657  8.9951
#> 
#> (2,2,2,.,.) = 
#>  -2.5830  2.6838
#>   1.3960 -2.1490
#>   3.6903 -4.3413
#>  -4.8245  1.9661
#>   2.4275 -0.8489
#> 
#> (3,2,2,.,.) = 
#>  -2.3136 -3.0029
#>   6.4553  2.9853
#>   1.0790  6.2816
#>  -2.3066  3.1048
#>   2.4425 -2.4980
#> 
#> (1,3,2,.,.) = 
#>   5.0888 -1.9831
#>  -3.3846 -3.1375
#>   0.4764  0.9040
#>  -0.2155 -3.9354
#>  -0.2403  9.0833
#> 
#> (2,3,2,.,.) = 
#>   3.2478  0.6904
#>   4.8131 -4.7587
#>   4.4327 -2.0460
#>  -0.1583 -1.1435
#>  -0.8185 -3.7916
#> 
#> (3,3,2,.,.) = 
#>   -4.8215  -5.7823
#>   -4.3197   2.5006
#>   -7.6764  10.5907
#>    5.7097   9.1375
#>   -6.5421   5.3465
#> 
#> (1,1,3,.,.) = 
#>   6.8857  1.6267
#>   0.6538 -2.9130
#>   4.2404 -5.3240
#>   5.6476  5.1706
#>   1.4154  2.3173
#> 
#> (2,1,3,.,.) = 
#>  -0.5342 -1.8588
#>  -6.3230 -2.3969
#>  -2.6561 -1.8841
#>   2.2272 -2.2146
#>  -7.2355 -9.8852
#> 
#> (3,1,3,.,.) = 
#>    6.9817  -1.9044
#>  -10.6065  -3.0032
#>    6.2727  10.5124
#>    2.0892  -0.2945
#>    1.4965   3.6209
#> 
#> (1,2,3,.,.) = 
#>  -2.7138  0.5805
#>   3.8351 -13.1060
#>  -2.0923  8.4118
#>   6.7002  2.8014
#>  -2.0329 -5.4260
#> 
#> (2,2,3,.,.) = 
#>  -4.7038  4.4041
#>   0.6079  6.9384
#>  -2.4581 -2.0337
#>  -2.4631  5.5419
#>  -2.4616 -2.3076
#> 
#> (3,2,3,.,.) = 
#>   1.0971  1.5684
#>   4.6985  3.6660
#>   1.7358  2.5220
#>  -2.9059 -3.8360
#>   8.5191 -0.6597
#> 
#> (1,3,3,.,.) = 
#>   1.2912 -4.1150
#>   5.9200 -9.2185
#>   2.7708 -11.2848
#>  -5.2274 -9.0754
#>  -5.9556 -8.3723
#> 
#> (2,3,3,.,.) = 
#>    3.6284  -9.7751
#>   -6.4936   0.4896
#>    1.0783   3.2584
#>   -1.2375   4.0910
#>    8.7281  11.1889
#> 
#> (3,3,3,.,.) = 
#>  -2.9392 -0.8457
#>  -5.1092 -3.4291
#>  -6.7226 -2.2762
#>  -6.7354  0.7375
#>  -3.6834 -10.3393
#> 
#> (1,1,4,.,.) = 
#>   -0.4641   5.1642
#>   -4.9807  -2.8183
#>   -0.9600   1.1617
#>   13.1615  -5.7607
#>   -5.8192   6.1872
#> 
#> (2,1,4,.,.) = 
#>   -0.6880  12.9307
#>   -8.0335  -1.1752
#>    6.2135   2.0547
#>    8.6658  -5.6305
#>   -7.1765  -4.2632
#> 
#> (3,1,4,.,.) = 
#>  -3.0057 -3.1364
#>   4.9857  4.4488
#>   1.3519  9.6340
#>  -9.3729  4.8508
#>   0.7863  3.1354
#> 
#> (1,2,4,.,.) = 
#>  -5.5472  0.7396
#>   5.5568 -2.7190
#>   2.1487  3.5182
#>   2.2908  5.8565
#>  -2.7511  0.1199
#> 
#> (2,2,4,.,.) = 
#>   0.1053 -9.7559
#>   2.6366  2.9785
#>   5.8546 -8.3557
#>   2.3308 -0.3229
#>  -0.1616 -2.8580
#> 
#> (3,2,4,.,.) = 
#>  -8.8217 -4.0197
#>  -1.3547  2.0932
#>   5.0738 -2.1672
#>  -1.6119  7.2245
#>  -1.4150 -0.6509
#> 
#> (1,3,4,.,.) = 
#>  -1.2559  2.1416
#>  -0.0231  4.1475
#>  -3.8181  2.3299
#>   3.7565  0.7653
#>   1.8683 -3.8644
#> 
#> (2,3,4,.,.) = 
#>    0.4839   6.4445
#>    3.1534  -8.8380
#>    3.7889   6.3242
#>    9.0306  11.1551
#>    5.7409  -1.1564
#> 
#> (3,3,4,.,.) = 
#>  -2.2988  8.8577
#>   4.7663 -5.2979
#>  -4.7160  9.3649
#>  -4.3578 -7.4509
#>  -4.3874  2.7093
#> 
#> (1,1,5,.,.) = 
#>  -0.4679  1.3350
#>  -1.4923 -5.4231
#>  -2.3291  0.3742
#>   2.6760 -0.3634
#>  -3.0602  8.7696
#> 
#> (2,1,5,.,.) = 
#>  -2.1765  4.7492
#>   7.2170  3.6409
#>  -6.5629 -2.2038
#>  -5.8527 -0.0878
#>  -4.5430  0.7514
#> 
#> (3,1,5,.,.) = 
#>   8.4563  6.2536
#>  -7.6490 -0.3940
#>  -2.2804  0.5806
#>  -9.5464  4.4593
#>  -5.9173 -3.0061
#> 
#> (1,2,5,.,.) = 
#>    5.1614  -5.6453
#>    4.8296 -15.8688
#>    1.8169  -0.8523
#>   10.0722  -2.0120
#>    0.8588   2.5174
#> 
#> (2,2,5,.,.) = 
#>  -6.1863 -1.2811
#>  -5.0314  3.6125
#>  -6.2021 -2.8083
#>  -0.3612 -3.9482
#>  -3.0259  0.6619
#> 
#> (3,2,5,.,.) = 
#>   5.1903  0.0372
#>  -1.9769  1.8278
#>   4.1834 -11.1886
#>  -3.8679 -4.4589
#>   2.2054 -3.4290
#> 
#> (1,3,5,.,.) = 
#>  -9.9394  5.2093
#>  -6.9233 -3.8417
#>   7.3800 -2.2016
#>   1.4532  4.9600
#>  -2.6033 -3.4185
#> 
#> (2,3,5,.,.) = 
#>  -6.9207  1.5530
#>  -3.9120  0.2946
#>  -0.9112  5.2904
#>   3.2102 -2.0272
#>   7.6431 -0.3506
#> 
#> (3,3,5,.,.) = 
#>   3.1340  6.7375
#>   1.4610  3.8716
#>   2.8714  3.5657
#>  -4.2452 -3.6816
#>   5.5423  2.7491
#> [ CPUFloatType{3,3,5,5,2} ]