Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   0.8455  3.4709
#>  -2.4137 -0.9884
#>  -2.6008 -1.7923
#> 
#> (2,.,.) = 
#>  -0.5982 -0.6358
#>  -1.8823 -7.8303
#>   2.0602  2.2420
#> 
#> (3,.,.) = 
#>   2.8272 -1.9673
#>   1.5493 -3.6119
#>  -5.6882 -5.0586
#> 
#> (4,.,.) = 
#>   0.9531  1.1699
#>  -2.2312 -2.1905
#>  -0.2325  1.2896
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>   1.0069  0.5094
#>  -1.2445 -3.6553
#>  -1.6153 -0.8298
#> 
#> (2,.,.) = 
#>  -0.0440  0.9717
#>   0.4192  0.7431
#>   0.5337  1.3898
#> 
#> (3,.,.) = 
#>   0.8295  0.2424
#>   0.8123  1.3551
#>  -2.5292 -2.5956
#> 
#> (4,.,.) = 
#>  -0.9469  1.7474
#>  -2.4007  0.5687
#>   1.0099  0.2434
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -10.3987   9.2806
#>   -0.8743  -5.5389
#>    3.6194  -6.8037
#>    5.4422  -9.9559
#>   -0.9513   6.9370
#> 
#> (2,1,1,.,.) = 
#>   5.3545 -1.9973
#>  -6.6644  3.1824
#>  -3.4250 -6.1113
#>  -2.2328 -1.0924
#>   4.2746 -1.2802
#> 
#> (3,1,1,.,.) = 
#>  -3.3479 -4.1803
#>  -0.1399 -0.4167
#>  -0.0731  7.3498
#>  -2.4250  0.8954
#>  -2.9822 -3.0903
#> 
#> (1,2,1,.,.) = 
#>  -0.0835  5.3480
#>  -11.9732  2.3115
#>  -0.6000 -2.2728
#>  -1.9932 -1.0839
#>   4.6444 -2.7698
#> 
#> (2,2,1,.,.) = 
#>  -5.8864  0.1611
#>  -0.5780  0.6334
#>   1.6233 -10.7673
#>  -3.4633  0.1536
#>   1.6124 -4.5111
#> 
#> (3,2,1,.,.) = 
#>  -6.2970  7.4450
#>  -1.7901  2.6024
#>  -3.9432  6.4693
#>  -6.9475  4.4887
#>  -1.3661  2.4789
#> 
#> (1,3,1,.,.) = 
#>   4.5478 -1.3884
#>  -0.5567 -2.1079
#>  -8.0256 -6.7543
#>   9.3980  3.8767
#>  -3.0583 -2.3708
#> 
#> (2,3,1,.,.) = 
#>   1.5625 -2.1804
#>   1.8681 -4.1282
#>  -0.9391 -1.0952
#>  -6.2609 -4.5528
#>  -1.3911 -3.0928
#> 
#> (3,3,1,.,.) = 
#>  -1.5018  2.5617
#>   1.8770  5.7833
#>   3.2065 -5.4844
#>   5.8919 -2.9841
#>   0.0986 -7.3148
#> 
#> (1,1,2,.,.) = 
#>   -2.2930   3.8534
#>   -1.2628   0.0417
#>   -3.0983   0.5378
#>   11.3584   1.3778
#>   -1.1354   7.5253
#> 
#> (2,1,2,.,.) = 
#>   8.6005 -9.8663
#>   0.3245  5.5781
#>  -1.1382 -8.6624
#>  -0.3124  5.2306
#>   6.8989  8.2791
#> 
#> (3,1,2,.,.) = 
#>  -3.8545  4.2541
#>   4.9377  0.2698
#>  -3.9948 -1.4339
#>  -0.4130 -0.6295
#>  -0.3998 -0.9005
#> 
#> (1,2,2,.,.) = 
#>   1.3566 -2.8808
#>   3.2078 -3.2750
#>  -0.3666 -1.7567
#>   4.7540 -0.1601
#>  -10.9092 -6.1929
#> 
#> (2,2,2,.,.) = 
#>   -0.9918  -0.4703
#>   -6.0328  -5.0102
#>   -0.7165   7.0621
#>   10.9361  -0.3281
#>    2.0596  -4.0446
#> 
#> (3,2,2,.,.) = 
#>   3.7675  2.8930
#>  -4.1257 -5.9714
#>  -3.6345  1.6325
#>   5.9324  6.2037
#>   2.1713 -6.3830
#> 
#> (1,3,2,.,.) = 
#>  -2.5373 -7.5182
#>   8.4228 -2.0305
#>  -4.4319  7.4726
#>  -5.8974 -0.8147
#>  -1.0620 -2.5875
#> 
#> (2,3,2,.,.) = 
#>    7.1027  -0.4865
#>    5.4568  10.1958
#>    2.9095   3.4585
#>    7.2248  -2.6724
#>   -0.9061   2.2704
#> 
#> (3,3,2,.,.) = 
#>  -5.7770 -1.4976
#>   1.2588  2.0766
#>   0.3093  0.2465
#>  -5.9511 -5.8122
#>   0.5149  3.5671
#> 
#> (1,1,3,.,.) = 
#>  -4.0564  6.3087
#>  -2.9401  0.0113
#>   6.7828 -2.6838
#>   8.5032  0.7547
#>  -12.0557 -5.6850
#> 
#> (2,1,3,.,.) = 
#>   -6.4382  -8.7142
#>   -1.3479   4.2252
#>   12.0990  -5.5005
#>    1.2326   4.9129
#>   -9.3990  -3.4121
#> 
#> (3,1,3,.,.) = 
#>  -4.8493 -1.0628
#>  -0.7883  6.7812
#>  -1.7633 -2.2719
#>   4.6444  3.0362
#>  -6.5751 -2.3059
#> 
#> (1,2,3,.,.) = 
#>    0.5909   2.5960
#>    0.0082   9.7080
#>   -5.2627  10.3521
#>   -2.2713  -1.8711
#>   -2.6710  -5.1583
#> 
#> (2,2,3,.,.) = 
#>  -8.2251 -2.3607
#>   1.5243  0.8157
#>  -0.7002  0.0770
#>  -4.1001 -9.9152
#>  -1.6806  0.3875
#> 
#> (3,2,3,.,.) = 
#>  -3.6101 -4.5752
#>  -3.6599  2.9927
#>  -5.4930 -7.4996
#>  -5.6068  0.2457
#>  -1.6843  1.7590
#> 
#> (1,3,3,.,.) = 
#>   -5.1161   0.9144
#>    5.2278   4.6059
#>   -3.5177   6.5155
#>   -3.6146  -2.0312
#>    2.1573  10.3089
#> 
#> (2,3,3,.,.) = 
#>   6.2586 -3.0435
#>   7.8546  5.5930
#>   2.6292  1.2789
#>  -0.2338 -11.0938
#>  -2.3085  9.8623
#> 
#> (3,3,3,.,.) = 
#>   11.2380   3.0080
#>   -6.3644  -2.2591
#>    4.6614   3.2337
#>   -4.6680  -3.6726
#>   -6.6602  -0.8858
#> 
#> (1,1,4,.,.) = 
#>   0.1824  3.9640
#>  -8.2745  0.9934
#>  -4.6958  5.5343
#>  -0.7827  0.1253
#>   5.3871  4.7025
#> 
#> (2,1,4,.,.) = 
#>  -0.6557  6.9496
#>  -2.0290  0.6028
#>  -0.0914  3.0465
#>   0.5733  4.0988
#>  -10.0674 -0.4331
#> 
#> (3,1,4,.,.) = 
#>  -2.2430  6.2471
#>   4.2747  6.1589
#>  -8.5753  4.4801
#>   4.0269 -5.1132
#>   0.1208 -4.0020
#> 
#> (1,2,4,.,.) = 
#>   1.4310  2.5301
#>  -4.1304  0.8054
#>   4.5141  2.9807
#>  -2.0671  3.3568
#>   0.5771 -2.1224
#> 
#> (2,2,4,.,.) = 
#>  -0.4243  1.6246
#>  -2.1024  1.2215
#>   4.8056  3.6178
#>  -8.7577 -1.6867
#>   7.7054  3.8193
#> 
#> (3,2,4,.,.) = 
#>   8.0580 -1.5962
#>   3.8897 -5.4042
#>   8.2054 -4.7755
#>  -0.4881  1.4315
#>  -5.4350 -1.3873
#> 
#> (1,3,4,.,.) = 
#>  -2.2105  3.4746
#>  -1.8669  5.5669
#>  -9.9577 -7.1559
#>  -1.8951  1.6661
#>   2.8890  0.0652
#> 
#> (2,3,4,.,.) = 
#>  -6.2812 -4.9340
#>  -0.2351 -1.3632
#>   3.8333 -8.6200
#>  -4.3622  0.8508
#>  -4.8687  2.6719
#> 
#> (3,3,4,.,.) = 
#>   -3.4617 -14.9408
#>    1.3259  -6.3439
#>   10.5979   5.2038
#>   -2.9457  -7.8692
#>   -6.1713  -1.8451
#> 
#> (1,1,5,.,.) = 
#>   3.5061  6.3626
#>   2.5397  1.0038
#>   2.2428 -0.2009
#>  -0.9183 -4.6854
#>  -1.8178  0.9064
#> 
#> (2,1,5,.,.) = 
#>  -11.2026   8.0634
#>   -4.4989  -1.3496
#>   -3.5813  -5.9882
#>   -3.5777  11.8980
#>    1.2940  -0.9847
#> 
#> (3,1,5,.,.) = 
#>   5.9286  0.0051
#>  -2.5625  3.2033
#>   2.9886 -4.8183
#>   1.5291  4.7471
#>  -4.6239  5.3382
#> 
#> (1,2,5,.,.) = 
#>   0.2594 -4.6492
#>  -6.9206 -1.4630
#>   2.5430 -4.6414
#>  -2.1870 -8.8772
#>   2.5493  1.3322
#> 
#> (2,2,5,.,.) = 
#>   1.1358 -5.0502
#>  -3.7082 -6.9809
#>   2.0125 -2.2275
#>   0.1435 -3.2747
#>  -4.9890 -0.2344
#> 
#> (3,2,5,.,.) = 
#>  -4.8482 -6.4918
#>   3.7547  2.5680
#>  -2.0964 -4.0245
#>  -6.3955 -6.1625
#>   4.5681 -0.1780
#> 
#> (1,3,5,.,.) = 
#>  -8.8694  1.4185
#>  -0.0220  3.1391
#>   8.1753  0.6879
#>  -1.7653 -5.1399
#>  -7.2749  2.3071
#> 
#> (2,3,5,.,.) = 
#>  -1.6329 -1.6560
#>   6.2776 -3.3200
#>  -8.5943 -3.1813
#>  -6.8484 -6.3933
#>  -3.1971 -3.7925
#> 
#> (3,3,5,.,.) = 
#>   5.6174 -8.8954
#>   1.0281  3.4356
#>  -3.6474  2.4562
#>  -6.1424 -7.3068
#>  -11.3471  7.1828
#> [ CPUFloatType{3,3,5,5,2} ]