Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# \dontrun{

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   1.6091  1.1177
#>   1.7387  0.5245
#>   1.3022 -7.5020
#> 
#> (2,.,.) = 
#>  -2.2896  1.5947
#>   2.3894 -3.7256
#>   2.4421 -6.0534
#> 
#> (3,.,.) = 
#>  -2.6684  0.2216
#>   4.6351  2.1387
#>   1.0104  2.1655
#> 
#> (4,.,.) = 
#>   0.6484  1.5756
#>   3.6644  0.9869
#>   0.8331  1.2714
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.6751  1.1274
#>   3.1069 -0.0189
#>   1.3969 -2.5296
#> 
#> (2,.,.) = 
#>   1.0646 -0.5105
#>   0.4540 -0.7223
#>   1.9041 -2.0146
#> 
#> (3,.,.) = 
#>   0.1455 -0.4577
#>   0.0800  1.3505
#>  -0.2407 -0.1386
#> 
#> (4,.,.) = 
#>   1.0741  0.9585
#>  -1.9022 -0.0848
#>  -1.7582 -2.8191
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -5.4725 -1.0036
#>  -0.4072 -9.9106
#>  -6.2423 -6.4070
#>  -1.9919 -0.2080
#>  -1.5095 -4.7941
#> 
#> (2,1,1,.,.) = 
#>   -1.3392   3.2072
#>   -2.8398   7.8139
#>    0.8047  -3.0656
#>   10.5509  -0.0158
#>    0.4441   1.4712
#> 
#> (3,1,1,.,.) = 
#>   5.8889 -0.8239
#>   0.6557 -4.5634
#>   3.6018 -5.1202
#>   6.7747 -1.8286
#>  -6.4471  3.7626
#> 
#> (1,2,1,.,.) = 
#>  -0.2606  1.6037
#>   1.2933  2.7796
#>  -1.2262 -8.3113
#>  -2.9285 -0.8387
#>  -8.8958  7.9346
#> 
#> (2,2,1,.,.) = 
#>   -5.8665  -5.8427
#>    3.5695  -0.4600
#>   10.6353  -4.9917
#>   -1.8417  -2.2605
#>    4.0205   3.3620
#> 
#> (3,2,1,.,.) = 
#>  -5.3450  5.1190
#>  -7.1003  3.1183
#>   2.1035 -0.7317
#>  -1.7222  6.8806
#>   1.1505  7.5300
#> 
#> (1,3,1,.,.) = 
#>  -9.1081  7.4756
#>   5.1849 -1.2153
#>  -1.6891 -2.1615
#>  -4.9926  0.1991
#>   5.4721 -2.1878
#> 
#> (2,3,1,.,.) = 
#>  -4.9459  0.0120
#>  -5.7250 -0.8370
#>   0.7157  1.3166
#>  -9.2864  4.1198
#>  -2.5485  6.5091
#> 
#> (3,3,1,.,.) = 
#>   2.9661 -4.1025
#>   5.0720  3.2858
#>  -1.6322 -9.3722
#>  -0.3661 -5.5639
#>   6.5081 -2.4129
#> 
#> (1,1,2,.,.) = 
#>   -2.3401   1.2643
#>    5.3967  -7.4990
#>    0.1779   3.4217
#>    6.0804   6.1065
#>   10.1270  -3.5843
#> 
#> (2,1,2,.,.) = 
#>   -9.0572  -1.9909
#>    4.2537  -0.0385
#>    5.9323  -1.1033
#>   -1.4235  11.0225
#>    0.3750   3.9369
#> 
#> (3,1,2,.,.) = 
#>   7.3848 -1.2410
#>  -1.4446  1.3323
#>  -1.4867  6.3887
#>  -1.3989  4.3918
#>  -1.6242  0.4001
#> 
#> (1,2,2,.,.) = 
#>   2.2515 -11.7348
#>  -2.9592 -0.4199
#>   6.0865 -2.5877
#>  -3.9066  0.8276
#>  -6.4707 -0.1932
#> 
#> (2,2,2,.,.) = 
#>   -3.5320 -15.9640
#>   -3.2020  -4.1973
#>    3.5907  -1.9048
#>    4.8187  10.5827
#>    2.0354   3.6088
#> 
#> (3,2,2,.,.) = 
#>  -1.5620  2.4318
#>  -4.4488 -3.6563
#>   1.4412  6.2894
#>   3.7142 -4.4966
#>  -1.9539  0.9322
#> 
#> (1,3,2,.,.) = 
#>   -2.5739  11.5350
#>   -5.6638  -1.9311
#>   -7.1538   1.8841
#>    9.9282  -5.0425
#>    4.0550 -13.0364
#> 
#> (2,3,2,.,.) = 
#>  -4.4776 -3.1815
#>   5.2017  7.8643
#>  -1.9960 -3.8433
#>   1.8730 -3.0428
#>   0.4528  1.8324
#> 
#> (3,3,2,.,.) = 
#>  -1.6393  3.0116
#>  -7.4910  4.6567
#>   3.9892 -1.2984
#>   2.3925  1.7542
#>  -0.4821  2.7867
#> 
#> (1,1,3,.,.) = 
#>  -1.9416  6.9300
#>  -5.9122  4.7269
#>  -6.1427 -3.6378
#>   4.8333 -7.6424
#>  -0.8626 -7.3429
#> 
#> (2,1,3,.,.) = 
#>   0.7169  7.3508
#>  -2.2415  0.6318
#>   8.5924 -3.5120
#>  -1.8660 -2.9021
#>   2.5242  6.2449
#> 
#> (3,1,3,.,.) = 
#>   3.2523 -9.4747
#>   6.7858  2.1141
#>  -3.1627 -8.3128
#>   8.1738  1.7333
#>  -12.5794  7.6793
#> 
#> (1,2,3,.,.) = 
#>   2.1388 -0.8053
#>  -7.5468 -4.9134
#>  -1.4956  0.7191
#>   4.5485  1.3213
#>   1.9775  3.2742
#> 
#> (2,2,3,.,.) = 
#>    2.7705  -3.2113
#>   -7.1385  10.5362
#>    2.1267  -3.4305
#>   -3.4051  -2.5411
#>   -2.7076   1.7485
#> 
#> (3,2,3,.,.) = 
#>    3.3394  10.9207
#>   -0.5970  -0.5034
#>    8.1799   3.0040
#>   -0.9163  -3.5497
#>   -3.1114  -5.1809
#> 
#> (1,3,3,.,.) = 
#>  -4.1027 -2.0095
#>   2.7722 -4.7394
#>  -3.5039 -1.9563
#>  -4.3085  8.7966
#>   2.7449 -4.6416
#> 
#> (2,3,3,.,.) = 
#>  -8.7689 -5.5072
#>   0.0556 -7.0627
#>   6.2366  2.7693
#>  -0.9913  0.2007
#>   2.4020  2.7007
#> 
#> (3,3,3,.,.) = 
#>    0.1972  -1.1722
#>    7.4460   3.4332
#>   -4.7450 -11.9086
#>   12.5079 -12.2708
#>   -2.9972  -0.8924
#> 
#> (1,1,4,.,.) = 
#>  -2.6641 -4.3296
#>   3.0735  1.2215
#>   0.2079 -4.6184
#>  -2.3720 -1.2896
#>  -3.9109 -9.9345
#> 
#> (2,1,4,.,.) = 
#>  -5.0131 -2.2123
#>   2.7328 -6.4654
#>  -3.3791 -1.4795
#>  -3.0048  4.2683
#>   8.6069  3.8774
#> 
#> (3,1,4,.,.) = 
#>   0.5565 -3.4152
#>   3.1825 -1.2100
#>  -5.6883 -4.5265
#>  -2.6019  6.2445
#>  -1.2380  0.0733
#> 
#> (1,2,4,.,.) = 
#>  -0.0619 -1.4112
#>   3.0482 -12.0317
#>   2.3142 -5.9550
#>   1.6332 -4.9998
#>   3.3110  3.2442
#> 
#> (2,2,4,.,.) = 
#>   5.7054 -4.5310
#>   4.0814  0.5770
#>  -7.4352  1.0886
#>  -4.4217 -0.5822
#>  -2.2274 -2.5406
#> 
#> (3,2,4,.,.) = 
#>   -6.0243   1.6214
#>    0.9687   2.5785
#>   -9.2330  12.6507
#>    0.1245  -3.6675
#>    2.2133  -0.1646
#> 
#> (1,3,4,.,.) = 
#>  -9.8819  5.0379
#>  -15.9787 -0.0044
#>  -3.3972 -4.1292
#>  -0.2631  3.1780
#>  -8.3382  2.3260
#> 
#> (2,3,4,.,.) = 
#>   2.6493  5.7085
#>  -0.8552 -6.9286
#>   8.8915  1.1790
#>  -8.0308 -2.3641
#>  -5.5359 -0.2864
#> 
#> (3,3,4,.,.) = 
#>   6.0509 -4.4619
#>   2.5174  1.0450
#>   3.2386 -2.9879
#>  -4.2332 -11.7932
#>   8.3415 -11.6524
#> 
#> (1,1,5,.,.) = 
#>  -2.6344  8.4594
#>  -9.3032 -0.0826
#>  -0.7539  0.4355
#>  -5.7352  6.4314
#>   1.3710  5.0478
#> 
#> (2,1,5,.,.) = 
#>   5.7266 -5.1236
#>  -3.1017 -0.2088
#>  -6.6610  5.8354
#>   3.1587  1.7131
#>   3.4059  0.6908
#> 
#> (3,1,5,.,.) = 
#>   12.7228   0.6750
#>   -4.4296   8.4994
#>    2.5491   1.5131
#>   -1.2545  -5.1292
#>    8.7953   3.5627
#> 
#> (1,2,5,.,.) = 
#>  -2.5561 -2.6800
#>   3.0704  6.9316
#>  -4.2936  0.3370
#>  -3.2507  9.0391
#>   7.8262 -2.8746
#> 
#> (2,2,5,.,.) = 
#>  -4.7915 -2.9852
#>   0.4093 -1.2396
#>   1.8851  0.8607
#>  -3.2552 -4.0731
#>   7.0195 -3.8280
#> 
#> (3,2,5,.,.) = 
#>  -6.9953 -12.4542
#>   7.8834  1.2176
#>   9.6941 -9.1636
#>  -15.3205  2.1126
#>   0.3537 -10.4525
#> 
#> (1,3,5,.,.) = 
#>  -10.7714   0.7571
#>   -0.7565  -0.4820
#>    5.6134  -8.9826
#>   -1.1373  -4.2921
#>    4.1096  -1.6137
#> 
#> (2,3,5,.,.) = 
#>   3.5382  6.4175
#>  -2.8167 -1.5735
#>   8.0241 -4.9140
#>   1.6143  7.2625
#>  -2.5749  4.8191
#> 
#> (3,3,5,.,.) = 
#>   9.8500 -0.4376
#>  -4.1255 -2.8707
#>  -4.6457 -0.7456
#>  -4.6735  2.8975
#>   2.7895  9.0651
#> [ CPUFloatType{3,3,5,5,2} ]

# }