Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   1.0920 -0.2723
#>   2.5079 -3.5369
#>   3.9757  6.2359
#> 
#> (2,.,.) = 
#>  -2.8858  3.5639
#>   3.8500 -0.5072
#>  -1.8635  0.1916
#> 
#> (3,.,.) = 
#>  -3.0585  5.1279
#>   5.2761 -2.9152
#>  -4.2408  3.1897
#> 
#> (4,.,.) = 
#>   1.0129  0.5696
#>  -2.1514  1.6195
#>  -0.7196 -0.1439
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.9599  2.2473
#>   2.3707 -1.3349
#>  -0.7120  2.3683
#> 
#> (2,.,.) = 
#>   0.2891 -2.3247
#>  -0.1604  1.3449
#>   1.9703  0.4756
#> 
#> (3,.,.) = 
#>  -0.0234  0.1805
#>   1.5214 -1.8911
#>   0.5795  2.3445
#> 
#> (4,.,.) = 
#>   1.7862 -0.3754
#>  -1.2237 -1.6558
#>   2.1380  1.0475
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -5.3842 -7.3058
#>  -4.7036  1.7868
#>   3.6433 -5.7833
#>  -7.1257  3.1612
#>   5.4576 -10.7452
#> 
#> (2,1,1,.,.) = 
#>   1.9066 -2.6735
#>  -2.3793 -1.8833
#>  -4.1669 -9.4328
#>  -1.5452 -4.3296
#>  -2.8799  3.0236
#> 
#> (3,1,1,.,.) = 
#>  -1.9799 -2.9364
#>   1.8837 -4.0660
#>   2.2460  9.1412
#>   4.8360  1.8790
#>   8.2191 -0.9493
#> 
#> (1,2,1,.,.) = 
#>  -2.8464 -6.1262
#>   1.5008 -3.5916
#>  -9.0714 -1.6527
#>   3.6008  4.1124
#>  -1.7525 -0.0524
#> 
#> (2,2,1,.,.) = 
#>    2.4616  -6.5774
#>    5.8510  -0.6112
#>   -6.0337 -10.1858
#>   -3.4539  10.1522
#>   -1.1480   5.3127
#> 
#> (3,2,1,.,.) = 
#>   -3.0077   0.8472
#>   -2.5513  -1.7432
#>    2.1522  11.7330
#>    5.1672  -2.7602
#>    0.6722  -3.7111
#> 
#> (1,3,1,.,.) = 
#>   1.6911  6.0023
#>   1.7944  3.3599
#>   0.7782  2.3066
#>  -0.1706  4.5088
#>  -4.8960 -2.6881
#> 
#> (2,3,1,.,.) = 
#>   -5.6775  10.6946
#>    2.9705  -5.4412
#>    2.8954  -4.2902
#>    5.2023  -5.7295
#>   -1.0059   2.7293
#> 
#> (3,3,1,.,.) = 
#>  -7.5152 -3.6646
#>   7.3902  0.8742
#>  -3.6710 -2.3228
#>  -6.0706  2.1242
#>  -8.0268 -8.1997
#> 
#> (1,1,2,.,.) = 
#>    9.0116   5.0715
#>   -1.6554  10.0222
#>   -1.6547  -3.7068
#>    9.3017  -3.3826
#>    0.6740   4.9413
#> 
#> (2,1,2,.,.) = 
#>   1.2520 -9.3973
#>   4.9044  8.0228
#>   9.3124  4.9603
#>   2.1008  0.1402
#>  -4.6243 -1.6374
#> 
#> (3,1,2,.,.) = 
#>   3.4692 -2.7515
#>  -3.1430 -8.6156
#>   0.1583 -1.1037
#>  -4.8931  2.7655
#>   3.0191  1.1167
#> 
#> (1,2,2,.,.) = 
#>   1.4082 -1.4383
#>   4.9914 -4.3034
#>   2.4342 -5.4049
#>   0.8043  4.9261
#>   3.1787  3.2756
#> 
#> (2,2,2,.,.) = 
#>   2.8804  5.1424
#>   7.6741 -0.3674
#>   0.1539  2.3517
#>  -5.1873  4.8209
#>   1.2302 -0.2535
#> 
#> (3,2,2,.,.) = 
#>  -1.8945 -4.4438
#>  -5.2978  3.3455
#>  -2.5686 -1.1495
#>  -5.6486 -6.6178
#>   5.2204 -10.0708
#> 
#> (1,3,2,.,.) = 
#>  -3.7238  5.6692
#>  -0.0273  5.8336
#>  -0.3208  1.5928
#>  -3.3531  8.1067
#>  -4.4226  8.8227
#> 
#> (2,3,2,.,.) = 
#>   -2.7298   8.3786
#>   -1.5491  -7.6572
#>   -0.9423  15.2057
#>    3.0282  -1.6581
#>   -9.2998  -0.9096
#> 
#> (3,3,2,.,.) = 
#>  -4.9790  0.7084
#>   6.0347 -1.4968
#>  -1.0464 -0.3666
#>   4.8172  2.6897
#>   3.2481  1.9993
#> 
#> (1,1,3,.,.) = 
#>    0.5588  -7.3006
#>   -3.8165   0.9295
#>    2.4281  -7.4655
#>   -1.7962  11.7810
#>    1.0619   5.5560
#> 
#> (2,1,3,.,.) = 
#>  -4.4915  1.5630
#>   1.0350  0.1867
#>   3.3101  3.7834
#>  -2.5516 -0.2421
#>  -5.8181  1.0682
#> 
#> (3,1,3,.,.) = 
#>   -6.8161  -1.7115
#>    3.4265  -0.4760
#>   -1.8472   2.0295
#>    8.4042  10.1805
#>    0.1020   2.8592
#> 
#> (1,2,3,.,.) = 
#>  -0.2913  3.1251
#>  -4.3514 -4.2158
#>   1.3685  5.5486
#>   2.7984  4.7961
#>  -6.1340  0.4552
#> 
#> (2,2,3,.,.) = 
#>  -2.5232 -4.2741
#>   6.9518  4.0747
#>   0.5612 -3.5502
#>  -9.4168  1.5225
#>  -0.9164 -7.5482
#> 
#> (3,2,3,.,.) = 
#>  -8.9133  6.7146
#>  -6.1157 -5.1770
#>  -4.8492 -0.0825
#>  -12.2994  4.2709
#>   4.9643  2.1670
#> 
#> (1,3,3,.,.) = 
#>  -2.1944 -4.8712
#>   2.3541  0.7245
#>   3.1709  1.8901
#>   2.0897  0.5381
#>  -3.7153  5.2608
#> 
#> (2,3,3,.,.) = 
#>  -5.9137 -1.7187
#>  -0.8520  2.2820
#>   0.9596  3.3001
#>   4.9131 -4.0544
#>   1.9730 -0.4293
#> 
#> (3,3,3,.,.) = 
#>   4.9170 -10.3231
#>   6.9721  1.5586
#>   1.6846 -6.4517
#>   1.4681 -1.4720
#>   8.9973 -3.7720
#> 
#> (1,1,4,.,.) = 
#>   1.4305  0.8573
#>  -10.7212 -10.0580
#>   1.5696 -5.5125
#>  -0.9686 -1.5139
#>  -2.6280  2.9523
#> 
#> (2,1,4,.,.) = 
#>   0.6863  1.2840
#>  -0.0793  4.3033
#>   4.8154 -2.0345
#>  -0.6611 -6.7750
#>  -2.6860 -3.5247
#> 
#> (3,1,4,.,.) = 
#>  -0.6174  3.2524
#>   3.3682 -4.5446
#>  -6.6384  5.9407
#>  -5.2130 -8.3305
#>   5.4363  6.4305
#> 
#> (1,2,4,.,.) = 
#>   1.1938  6.2629
#>  -1.2848 -0.7175
#>   3.9911  2.4203
#>   7.7121 -5.2472
#>  -13.5481 -6.4250
#> 
#> (2,2,4,.,.) = 
#>   1.7537  4.9987
#>  -5.9977  0.7866
#>   2.0596  0.6178
#>   2.9428 -10.3228
#>  -13.0765 -3.1040
#> 
#> (3,2,4,.,.) = 
#>   2.3315  9.3138
#>  -1.0706  1.7959
#>  -4.5510 -0.4155
#>  -0.7446  9.8425
#>  -5.6620  1.0830
#> 
#> (1,3,4,.,.) = 
#>  -7.8228 -8.6660
#>  -7.8007  6.7628
#>  -0.2990 -2.0439
#>   9.9119 -0.2287
#>   4.9641 -6.6724
#> 
#> (2,3,4,.,.) = 
#>    3.1289   6.9922
#>  -10.3613  -0.9191
#>   -8.0509   7.5990
#>   -0.4232  16.3819
#>   -4.4819   0.4685
#> 
#> (3,3,4,.,.) = 
#>  -7.0211 -1.8959
#>  -5.1470 -1.4139
#>   5.3861 -1.5495
#>  -0.5892  1.3239
#>  -4.5516  2.4334
#> 
#> (1,1,5,.,.) = 
#>   3.3629  3.2928
#>   1.3041  6.3400
#>  -10.5422 -5.0987
#>   6.0147 -3.0667
#>   9.4220  0.9036
#> 
#> (2,1,5,.,.) = 
#>  -0.4931  0.5443
#>  -1.6934  3.5712
#>  -4.1114  0.0897
#>  -2.7470  7.1554
#>  -5.3287 -1.1787
#> 
#> (3,1,5,.,.) = 
#>   11.4850   6.4792
#>    2.5458   2.6333
#>    4.2740  -5.0775
#>   -7.5269   3.8616
#>   -0.0230  -1.8652
#> 
#> (1,2,5,.,.) = 
#>   1.8676 -10.7473
#>   1.3769  3.9589
#>  -3.0087 -6.0852
#>   1.3448  1.5546
#>   0.7502 -2.8319
#> 
#> (2,2,5,.,.) = 
#>  -5.4415 -11.4466
#>   8.2112 -0.3075
#>  -1.9444  8.7701
#>   3.0985  2.4828
#>   2.6104 -5.4673
#> 
#> (3,2,5,.,.) = 
#>   5.3970  0.8321
#>   4.9694  3.9270
#>   2.4252 -0.9356
#>  -7.8651  0.1915
#>   3.5595 -1.1118
#> 
#> (1,3,5,.,.) = 
#>  -2.1818  4.2104
#>  -2.5053 -0.6936
#>   6.6726  8.2551
#>  -4.9805 -4.6143
#>   4.1919 -2.9348
#> 
#> (2,3,5,.,.) = 
#>  -6.6584 -7.5216
#>   9.8489  3.8002
#>  -5.2584 -2.5495
#>  -4.9161  6.8147
#>  -9.5074 -1.0884
#> 
#> (3,3,5,.,.) = 
#>   0.9202 -4.1401
#>  -0.2764  8.0664
#>  -6.1060  2.3893
#>   6.1835 -0.6001
#>   0.5580 -3.8095
#> [ CPUFloatType{3,3,5,5,2} ]