Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.7319 -2.2375
#>   2.4296  2.9927
#>  -1.0356 -2.4704
#> 
#> (2,.,.) = 
#>   2.9782  2.0781
#>  -4.3604 -4.6423
#>   2.6659  2.2724
#> 
#> (3,.,.) = 
#>   4.8448  4.3647
#>   5.7090  0.2738
#>  -4.9717  1.0800
#> 
#> (4,.,.) = 
#>   8.0669  0.5732
#>  -4.6046  3.9773
#>   3.3943  4.5214
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>   3.7895  1.1946
#>  -0.2066  0.6504
#>   0.0132  1.3509
#> 
#> (2,.,.) = 
#>  -1.7704 -2.9227
#>   1.3351  0.7408
#>   1.5463 -1.0697
#> 
#> (3,.,.) = 
#>  -1.7330 -0.1311
#>   4.2759  0.9829
#>  -3.0169 -2.0461
#> 
#> (4,.,.) = 
#>  -1.0180 -0.3784
#>  -2.9748  0.6187
#>   0.4218 -0.7055
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -3.0270 -12.7802
#>   1.4201  8.5574
#>  -2.3437  2.1055
#>   5.7898  6.3273
#>   6.9509 -4.3546
#> 
#> (2,1,1,.,.) = 
#>    7.5813  -4.1829
#>   -3.7127   2.0008
#>   10.1061  -0.6909
#>    1.8361  -5.2007
#>   -1.8122 -10.5429
#> 
#> (3,1,1,.,.) = 
#>   -5.7053   1.8859
#>   -5.6316  10.3033
#>    1.4941   1.3529
#>   -5.9251   3.0446
#>    3.8566  -7.6872
#> 
#> (1,2,1,.,.) = 
#>  -2.5986  5.6254
#>  -2.5646  2.6349
#>   3.9251 -1.7533
#>  -4.3010 -6.3840
#>  -3.0057  1.3619
#> 
#> (2,2,1,.,.) = 
#>  -3.0055 -10.2855
#>   0.4977  2.5084
#>  -0.8988  0.9638
#>  -0.3201  1.8817
#>   3.3391 -4.6866
#> 
#> (3,2,1,.,.) = 
#>   4.0350 -2.4348
#>   7.7836 -7.7825
#>  -0.0435 -10.7669
#>  -8.8972  0.8771
#>   0.9256 -1.1031
#> 
#> (1,3,1,.,.) = 
#>   4.0732 -1.7324
#>  -2.5688 -0.6805
#>  -12.1112 -2.3649
#>   3.3661  5.9049
#>   2.5260  2.5422
#> 
#> (2,3,1,.,.) = 
#>   0.7891  2.2631
#>   5.3054  1.9675
#>   0.9267 -9.5757
#>  -6.9103 -4.5743
#>  -2.3402 -9.0725
#> 
#> (3,3,1,.,.) = 
#>   5.9459  2.7989
#>   2.2311  3.4644
#>  -2.0887  8.1477
#>   9.3009 -3.0210
#>   1.4427 -2.5925
#> 
#> (1,1,2,.,.) = 
#>   4.3494 -3.5881
#>   1.5450 -6.6832
#>   3.4103 -1.9343
#>   1.9800 -5.3762
#>   4.0015  0.0777
#> 
#> (2,1,2,.,.) = 
#>    2.5402   4.2167
#>    4.2702   3.2909
#>   -4.5585   2.1322
#>    2.1025  -4.6064
#>   -0.1101  11.9281
#> 
#> (3,1,2,.,.) = 
#>    8.4416  -0.9108
#>   12.0541   5.8229
#>    2.0444   0.4531
#>    0.0786  -1.4094
#>   -2.6686  -2.2156
#> 
#> (1,2,2,.,.) = 
#>    0.5957   6.4551
#>    0.8756   3.0904
#>   -4.0919   5.9729
#>   -5.6541   6.5868
#>    4.9072  12.1454
#> 
#> (2,2,2,.,.) = 
#>   0.8240 -9.2634
#>   0.6104 -6.8459
#>  -2.9687  2.5982
#>  -8.7795  0.8351
#>   5.2345  5.7019
#> 
#> (3,2,2,.,.) = 
#>   1.3295  2.3046
#>  -5.0923  1.6443
#>   6.3156 -3.8553
#>   0.3386 -6.8998
#>  -0.0026 -3.5591
#> 
#> (1,3,2,.,.) = 
#>  -3.2091 -2.0388
#>   2.4838  8.6878
#>   1.8786 -5.2886
#>   5.2213 -0.3809
#>   4.0257  3.8770
#> 
#> (2,3,2,.,.) = 
#>    3.0987  -2.2104
#>   -1.4231   4.1041
#>    7.6174   2.6902
#>    2.4008  -2.3201
#>   10.3434   5.1905
#> 
#> (3,3,2,.,.) = 
#>    1.3777  -0.9117
#>   10.3789   8.1560
#>   -4.3676  -0.2411
#>    0.9966  11.7826
#>    1.0074   2.1776
#> 
#> (1,1,3,.,.) = 
#>   8.5069  7.2849
#>   3.0725  7.7459
#>  -8.2301  0.5564
#>   3.7580 -6.8937
#>   7.8104  0.8826
#> 
#> (2,1,3,.,.) = 
#>   8.3870  8.7746
#>  -6.6292 -1.4022
#>   3.0195 -5.6831
#>  -9.9837  4.8501
#>   4.3797  0.6527
#> 
#> (3,1,3,.,.) = 
#>   -0.7051  -5.7201
#>   -5.1145   1.4312
#>   10.5566  -3.4431
#>   -1.5513 -14.1361
#>   -3.8271  -9.7999
#> 
#> (1,2,3,.,.) = 
#>    9.8515 -10.8005
#>   -6.9267  -2.1440
#>    3.8696   4.5214
#>   10.3227   4.4173
#>    9.7854  -7.3236
#> 
#> (2,2,3,.,.) = 
#>   -3.8411   1.8637
#>   -0.3704  -2.5988
#>    2.5957  -5.6097
#>   11.9757   1.2760
#>   -2.6745   0.2280
#> 
#> (3,2,3,.,.) = 
#>   7.0826 -10.6047
#>  -6.7525 -2.1293
#>   3.5464  6.5085
#>   3.0444  0.2479
#>  -7.4031 -6.9875
#> 
#> (1,3,3,.,.) = 
#>   -6.7146   2.5714
#>   11.1867  -6.9506
#>    0.2358   6.0853
#>   -3.6169   9.6309
#>    4.0797   7.6201
#> 
#> (2,3,3,.,.) = 
#>   7.2283 -4.0974
#>  -4.8273  3.7157
#>  -7.7398  6.4573
#>   0.9614 -0.9049
#>  -3.7047 -1.9025
#> 
#> (3,3,3,.,.) = 
#>   3.2009 -0.0338
#>   4.8916 -2.4113
#>   1.8366  5.3117
#>  -0.1205  3.5910
#>  -7.8022  0.0616
#> 
#> (1,1,4,.,.) = 
#>   4.1809 -4.1511
#>  -6.0050  7.4605
#>  -0.9812 -1.3206
#>   3.9098 -0.9893
#>  -8.2212 -6.6410
#> 
#> (2,1,4,.,.) = 
#>  -4.8466  1.6806
#>   2.9257 -1.8844
#>  -2.9106  4.8837
#>  -1.4043  5.8489
#>  -3.7670  5.2999
#> 
#> (3,1,4,.,.) = 
#>  -4.6434 -3.6964
#>   5.3929  3.9080
#>  -0.4919 -10.6924
#>   1.3785  7.2231
#>   4.2559  0.2260
#> 
#> (1,2,4,.,.) = 
#>   5.2726  1.2411
#>   1.7680 -8.1225
#>   2.0711  4.4615
#>   0.0795  2.6462
#>   7.5429 -9.7207
#> 
#> (2,2,4,.,.) = 
#>   2.9065  6.2854
#>   9.2547 -0.1567
#>   5.9428  3.0647
#>   1.5477 -2.3047
#>  -6.0802 -1.1224
#> 
#> (3,2,4,.,.) = 
#>    7.0135   7.5734
#>   -2.3947  10.3437
#>   -0.4880  -2.2032
#>   -1.3537   0.4868
#>   -5.3791   0.9910
#> 
#> (1,3,4,.,.) = 
#>   -2.7781  -3.4093
#>    0.2804  -8.2816
#>   11.7806  -1.6807
#>    3.0094   1.5970
#>   -9.4527  -8.6389
#> 
#> (2,3,4,.,.) = 
#>   3.4638 -2.5025
#>  -3.3911  6.4217
#>  -4.1976  0.7185
#>   1.0376  1.0986
#>  -4.2349 -4.6659
#> 
#> (3,3,4,.,.) = 
#>  -2.5362  1.6530
#>  -2.2721 -2.6943
#>   1.6465  1.8361
#>  -2.8789  7.1011
#>  -6.5450 -1.2362
#> 
#> (1,1,5,.,.) = 
#>  -1.2434 -0.0768
#>  -2.5326 -0.1825
#>   0.0347 -2.8052
#>   1.3882  4.8841
#>  -12.2493 -0.7069
#> 
#> (2,1,5,.,.) = 
#>   0.9731 -1.8823
#>  -0.1186 -2.3122
#>  -3.6737 -1.7161
#>  -6.4331 -6.0290
#>   1.3208  9.0370
#> 
#> (3,1,5,.,.) = 
#>   0.6791  0.0116
#>  -3.1466  0.9142
#>   3.8885 -1.0943
#>   4.8344 -4.8064
#>   4.1595 -1.8651
#> 
#> (1,2,5,.,.) = 
#>  -4.5012 -3.5562
#>  -3.8449 -0.2910
#>  -4.8218  2.8949
#>   8.4354 -5.2424
#>   1.8734  3.5961
#> 
#> (2,2,5,.,.) = 
#>   2.9108  3.5408
#>  -5.6800 -5.6701
#>   4.6722  0.1537
#>   9.4756 -2.9326
#>  -8.1583 -1.9940
#> 
#> (3,2,5,.,.) = 
#>    5.1473  -6.2254
#>    3.0489  -5.7761
#>   -4.8542  11.7846
#>  -11.5273   6.9082
#>    0.4646   3.7184
#> 
#> (1,3,5,.,.) = 
#>  -6.9062 -6.0254
#>  -2.9201 -2.5649
#>  -3.2596  5.4785
#>  -1.6510 -1.8427
#>   2.9466 -4.3485
#> 
#> (2,3,5,.,.) = 
#>   0.6690 -1.8485
#>   5.8763  2.3357
#>  -4.5015 -0.3753
#>   0.4528 -4.4912
#>  -1.6366 -4.3262
#> 
#> (3,3,5,.,.) = 
#>   1.8305 -3.1071
#>   0.9258  5.2358
#>  -4.3460 -4.0494
#>  -3.1750 -2.7844
#>   1.8932 -1.3865
#> [ CPUFloatType{3,3,5,5,2} ]