Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   -0.6237  11.1240
#>   -2.9024   3.9248
#>    0.9614   1.1158
#>    4.8762  -7.3797
#>  -11.2127   8.7440
#> 
#> (2,1,1,.,.) = 
#>   4.6478  0.7745
#>  -3.6287  8.6404
#>  -6.8821  2.6425
#>   8.1010 -4.4391
#>  -3.3068 -1.6102
#> 
#> (3,1,1,.,.) = 
#>    3.4637   9.3338
#>   -1.5727   2.7785
#>   -5.9445  -0.7541
#>   12.7314   3.5974
#>    4.6867  -3.8250
#> 
#> (1,2,1,.,.) = 
#>   -0.5954   4.3975
#>   -1.5096   3.8106
#>    5.5417  -3.2488
#>    0.2471   2.5054
#>   10.5497  -3.5451
#> 
#> (2,2,1,.,.) = 
#>  -3.2585 -9.5019
#>   1.4039  0.5617
#>  -1.6008 -4.4729
#>   5.2951  3.4503
#>   3.7162 -6.5419
#> 
#> (3,2,1,.,.) = 
#>   5.0617  7.7736
#>  -0.4432  1.6148
#>   2.7415 -3.3881
#>  -0.0691  5.2788
#>   2.9172  5.1361
#> 
#> (1,3,1,.,.) = 
#>  -4.7322 -1.3954
#>  -6.2138 -0.8846
#>  -6.2453 -5.9470
#>  -1.5404 -0.6660
#>  -0.2902 -1.4784
#> 
#> (2,3,1,.,.) = 
#>   -1.3700  -0.3495
#>    0.8277  -0.4278
#>   -3.3411   2.8361
#>   -2.2102  10.9602
#>   -2.4252  -8.4254
#> 
#> (3,3,1,.,.) = 
#>   3.8574  2.3426
#>  -6.9791 -4.6865
#>   9.2564 -7.2672
#>   5.5261  6.9735
#>   1.1010 -0.9819
#> 
#> (1,1,2,.,.) = 
#>   7.3214 -0.8657
#>   5.2632 -10.6552
#>  -9.0971 -2.9376
#>  -5.7490 -4.5256
#>   2.5403  4.0484
#> 
#> (2,1,2,.,.) = 
#>   6.5555  0.2084
#>   5.5978  0.1425
#>  -5.9039  1.2875
#>   5.8817 -0.3736
#>  -4.1912  0.7234
#> 
#> (3,1,2,.,.) = 
#>   0.6209  4.6246
#>  -1.7146 -2.3267
#>   7.8222 -6.6382
#>  -2.2250 -0.1006
#>  -0.1622  4.2237
#> 
#> (1,2,2,.,.) = 
#>   2.6396 -0.2409
#>   6.1416 -3.1177
#>  -5.5715  0.9069
#>   3.9293 -8.4025
#>  -4.7713 -1.8838
#> 
#> (2,2,2,.,.) = 
#>   4.4548  1.7675
#>  -4.6577  6.3155
#>  -3.6555  3.4786
#>  -5.0933  0.8499
#>  -5.9548 -1.3470
#> 
#> (3,2,2,.,.) = 
#>   3.7026  7.1005
#>  -10.4819  3.3709
#>   7.6422  3.8034
#>   1.3818 -9.0053
#>  -0.2780 -1.4319
#> 
#> (1,3,2,.,.) = 
#>   10.8958   1.0035
#>    9.9232   1.8652
#>    5.6906   4.3295
#>   -1.0827   2.9208
#>   -4.7428  -0.9122
#> 
#> (2,3,2,.,.) = 
#>   0.0704  1.7448
#>  -2.8207  0.2934
#>   3.4910 -4.0675
#>   3.8336 -5.5547
#>   7.7508  3.2039
#> 
#> (3,3,2,.,.) = 
#>  -1.3815 -7.5524
#>  -8.5848 -1.7580
#>  -4.4030  2.8512
#>   4.5375 -6.4783
#>  -6.7308 -6.7774
#> 
#> (1,1,3,.,.) = 
#>   4.5218 -5.3473
#>  -0.8311 -2.5596
#>   4.1558  2.0263
#>   3.0365  0.6949
#>   5.4967  2.3332
#> 
#> (2,1,3,.,.) = 
#>   3.5842  1.5343
#>  -1.1246 -1.7734
#>   3.6410 -8.1156
#>  -3.7644  6.3303
#>   4.4594 -7.4386
#> 
#> (3,1,3,.,.) = 
#>  -1.9088  3.4287
#>   8.0339 -0.0081
#>  -0.6776  4.8811
#>  -3.0000  1.4812
#>   0.4720 -3.9359
#> 
#> (1,2,3,.,.) = 
#>   8.9398 -1.3440
#>  -5.3892  3.7728
#>  -7.2133 -6.8858
#>  -1.8656 -1.8717
#>  -6.9839  6.3163
#> 
#> (2,2,3,.,.) = 
#>  -3.9451 -1.8309
#>  -4.8392 -1.8213
#>   3.9511  1.3351
#>  -4.0465  6.3634
#>   4.7880 -6.1240
#> 
#> (3,2,3,.,.) = 
#>   0.1236 -7.1395
#>  -8.0469 -1.5676
#>  -2.7822  0.9380
#>  -1.1440 -10.0607
#>  -5.9925  4.1269
#> 
#> (1,3,3,.,.) = 
#>   0.3911  0.3196
#>   3.0803 -4.6720
#>   3.5518  3.2623
#>   3.5829 -7.8312
#>  -3.1654 -3.7237
#> 
#> (2,3,3,.,.) = 
#>  -1.1693  8.9020
#>   2.7565 -2.3431
#>  -4.8860  7.4424
#>  -3.0172  4.9891
#>  -1.2997  5.7028
#> 
#> (3,3,3,.,.) = 
#>   -1.2071  13.5275
#>   -4.9776   1.4373
#>    8.8954   0.3671
#>   -2.8415  -7.6662
#>   -5.6262  -1.8897
#> 
#> (1,1,4,.,.) = 
#>   -2.4514   3.1007
#>   -7.9586  -1.3171
#>  -10.5946  -0.0578
#>   14.0214   4.5042
#>   -4.1099   1.2902
#> 
#> (2,1,4,.,.) = 
#>   0.8533 -1.4448
#>   4.4863 -2.3259
#>  -1.4015 -3.7351
#>   6.1828 -2.8592
#>  -1.8771  0.0266
#> 
#> (3,1,4,.,.) = 
#>  -0.8520 -2.2815
#>  -1.2792 -3.9903
#>  -14.5931 -4.0195
#>   4.2526  8.8196
#>   4.9681 -1.0271
#> 
#> (1,2,4,.,.) = 
#>    2.1900  -0.3117
#>    0.3076  15.2198
#>    0.7374  10.8459
#>   -4.4270  -6.2326
#>   -5.4171   7.2504
#> 
#> (2,2,4,.,.) = 
#>  -0.4042 -1.3235
#>  -4.0560 -3.9442
#>   2.5276  7.9577
#>   0.8329  1.3821
#>   0.3406  4.2565
#> 
#> (3,2,4,.,.) = 
#>  -5.5086 -14.2807
#>   1.7364 -3.1041
#>  -4.4011  9.3696
#>   8.2298 -0.5459
#>  -0.2637 -4.5050
#> 
#> (1,3,4,.,.) = 
#>   10.8423  -4.9814
#>    0.9733   0.5843
#>   -1.9394  -4.8631
#>   -2.8202  -4.1081
#>   -1.9058  -8.7359
#> 
#> (2,3,4,.,.) = 
#>   2.8039  2.0462
#>   0.9798 -0.7275
#>  -3.3579  7.8153
#>   5.9547  2.8450
#>   4.3205  2.1470
#> 
#> (3,3,4,.,.) = 
#>   9.8274  3.7765
#>   2.5813 -2.4610
#>  -0.3217 -10.1219
#>  -6.4127 -6.6547
#>   7.7007 -2.1778
#> 
#> (1,1,5,.,.) = 
#>    8.3611  -3.7002
#>  -10.3652  -4.8776
#>    2.2009   5.3398
#>    2.0239  -4.4912
#>   -0.2720  10.7965
#> 
#> (2,1,5,.,.) = 
#>   7.5661 -4.1393
#>  -2.3611  6.5618
#>  -10.5304  4.7209
#>   1.1810 -1.5315
#>   3.3532 -1.7273
#> 
#> (3,1,5,.,.) = 
#>   -5.5761   6.9648
#>   13.8696   0.2313
#>  -10.0644  -8.0058
#>   -0.9818   5.0198
#>   -4.9157  -1.7264
#> 
#> (1,2,5,.,.) = 
#>   0.8571 -1.0611
#>  -9.5913  2.8175
#>  -7.0851 -2.7899
#>  -7.2656  3.7659
#>  -1.3606  4.6802
#> 
#> (2,2,5,.,.) = 
#>  -3.1693 -2.1234
#>  -1.7774  4.9965
#>   3.0103 -5.4805
#>  -0.5224 -1.1890
#>   0.4450  2.9462
#> 
#> (3,2,5,.,.) = 
#>  -6.5190 -3.2037
#>   0.4573 -3.5337
#>   6.9170 -14.2690
#>  -4.1715 -0.1801
#>  -8.8619 -1.0151
#> 
#> (1,3,5,.,.) = 
#>  -1.0744 -2.5077
#>  -2.7232 -6.3829
#>  -9.4249  5.2446
#>  -1.4348 -0.4710
#>   3.3098  1.9990
#> 
#> (2,3,5,.,.) = 
#>  -2.7469  1.2545
#>   2.5361 -6.3646
#>   0.1712 -0.7867
#>  -5.9268  3.4458
#>  -0.2139 -9.2211
#> 
#> (3,3,5,.,.) = 
#>   0.8565 -14.5808
#>  -2.0871  0.8911
#>   3.2479  4.0964
#>   4.3496 -0.7272
#>   8.0389 -5.8174
#> [ CPUFloatType{3,3,5,5,2} ]