Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   2.1112 -4.1876
#>   9.0539 -0.7684
#>  -3.3602 -6.6980
#>  -3.8372  4.3760
#>   2.1318  2.8328
#> 
#> (2,1,1,.,.) = 
#>  -3.8660  9.2121
#>   4.0887 -0.4134
#>  -4.6112  2.0372
#>   8.7528  5.4361
#>  -0.4073 -3.2771
#> 
#> (3,1,1,.,.) = 
#>  -11.7211  -0.6949
#>    3.2481  -6.2415
#>    1.2418   0.7189
#>    2.5031  -2.6676
#>   -0.2124   1.8314
#> 
#> (1,2,1,.,.) = 
#>   -1.1248  -0.2299
#>    1.7505   4.6392
#>   -5.2587  -0.5909
#>   -2.8888   0.2741
#>   -1.5589  12.0204
#> 
#> (2,2,1,.,.) = 
#>   -1.5638   6.6315
#>    2.8193   0.5542
#>    3.7464  -4.0393
#>   -0.2503  10.7857
#>  -10.6686   3.5195
#> 
#> (3,2,1,.,.) = 
#>   7.4754  9.3271
#>  -6.1667 -8.6362
#>   1.1637  5.9735
#>  -3.2823 -5.0654
#>  -6.6789 -5.9552
#> 
#> (1,3,1,.,.) = 
#>   6.2255  5.2914
#>  -3.1147 -9.1653
#>   4.4511  7.9658
#>  -2.0946  4.0487
#>  -6.1687  0.1991
#> 
#> (2,3,1,.,.) = 
#>   0.8664  0.3790
#>   2.5459 -6.9351
#>   0.8175 -4.5158
#>   3.6217  4.1695
#>   5.9312  7.1426
#> 
#> (3,3,1,.,.) = 
#>  -1.1202  1.0072
#>   1.0848 -8.0716
#>  -6.1188  3.0874
#>  -1.2835 -0.2749
#>   3.2995 -13.0627
#> 
#> (1,1,2,.,.) = 
#>  -6.8476 -2.0641
#>  -5.9904  4.9265
#>   0.5894 -14.4109
#>  -2.1048 -4.5217
#>   4.7026 -10.5779
#> 
#> (2,1,2,.,.) = 
#>    0.7832  -0.6694
#>   -3.0682  10.9361
#>   11.5658  -3.7508
#>    2.8428   9.1761
#>   -2.3998   4.1248
#> 
#> (3,1,2,.,.) = 
#>   1.0646  1.2157
#>  -0.6122  3.0589
#>   1.5070  1.9241
#>   5.4853  0.6845
#>  -1.9869 -3.9561
#> 
#> (1,2,2,.,.) = 
#>   -1.9126  -2.2446
#>    0.2187  10.6384
#>   -4.6431  -2.6122
#>    4.4661  -2.3179
#>    9.8271  -3.9473
#> 
#> (2,2,2,.,.) = 
#>  -3.4948  2.7012
#>  -1.4750  2.5726
#>   0.6487 -2.4248
#>  -7.9717  6.9759
#>   2.6088 -3.2346
#> 
#> (3,2,2,.,.) = 
#>  -4.6681  6.0932
#>  -3.3548 -4.0221
#>  -6.4799  4.9227
#>   1.6389 -0.3977
#>   7.0018 -1.1551
#> 
#> (1,3,2,.,.) = 
#>   -1.1084   5.8721
#>   -8.0657  -1.6099
#>   -4.8355  -0.7776
#>   -6.7260  10.8151
#>   10.9844  -0.1199
#> 
#> (2,3,2,.,.) = 
#>   0.3810  2.4896
#>  -6.7450  1.6632
#>   6.5421 -7.0797
#>  -4.1964 -5.5096
#>   1.0053  0.3290
#> 
#> (3,3,2,.,.) = 
#>  -2.4359 -3.1135
#>  -1.4993 -5.9281
#>   6.5173  7.9458
#>  -5.5326  0.4161
#>   0.6151 -2.9904
#> 
#> (1,1,3,.,.) = 
#>    4.8682  -4.1154
#>   -1.0644   0.3901
#>    2.3832   0.7723
#>    5.6018  -0.0959
#>   11.9678  12.8629
#> 
#> (2,1,3,.,.) = 
#>   5.8731  3.9244
#>  -2.6767 -3.7312
#>   9.4703 -0.5039
#>   3.3932 -1.5033
#>   0.5638 -5.0444
#> 
#> (3,1,3,.,.) = 
#>   4.9481  1.2354
#>  -0.4381  9.7790
#>  -6.4056  2.2659
#>  -6.1752  6.0914
#>  -2.7765 -7.5988
#> 
#> (1,2,3,.,.) = 
#>  -5.1882  3.7246
#>  -2.3062 -6.5209
#>   5.5984  7.8336
#>   1.0851 -6.7997
#>  -2.1747 -0.0115
#> 
#> (2,2,3,.,.) = 
#>   1.8408  8.3407
#>   1.6358  0.4085
#>   8.5524 -4.7476
#>  -5.4625 -0.4488
#>   1.4273  4.0275
#> 
#> (3,2,3,.,.) = 
#>   10.6421  -4.3512
#>   -1.0465  -1.5097
#>    2.9777   1.7065
#>   -9.8611   7.2358
#>    3.4448  -4.9188
#> 
#> (1,3,3,.,.) = 
#>  -2.1104  5.1731
#>  -1.7900 -7.7720
#>   0.1296  3.3414
#>  -4.6332 -2.0916
#>   2.4158  0.8121
#> 
#> (2,3,3,.,.) = 
#>   2.8492  5.0554
#>   1.7292  0.4023
#>   2.8614  0.1382
#>  -3.7582  0.1076
#>  -4.2009 -6.5711
#> 
#> (3,3,3,.,.) = 
#>   1.0870 -2.5169
#>   3.2648 -2.7512
#>  -3.8948  0.4316
#>   2.1210 -3.3306
#>  -0.6145  1.4705
#> 
#> (1,1,4,.,.) = 
#>   7.9321  3.5210
#>   1.6832 -1.9867
#>   5.4036 -3.9730
#>   2.9738 -5.7481
#>  -2.6489 -4.8766
#> 
#> (2,1,4,.,.) = 
#>   1.6424  3.1836
#>   4.5096  1.4822
#>  -3.5779  2.3471
#>   6.2285  5.4668
#>   0.6169 -1.4824
#> 
#> (3,1,4,.,.) = 
#>    4.6550   2.1552
#>   11.3668   2.2088
#>    2.3912   8.5285
#>    2.8511  -4.5633
#>   -2.3768   3.2482
#> 
#> (1,2,4,.,.) = 
#>   1.8475  4.6616
#>  -2.6992  5.8692
#>   1.2643  3.6892
#>   7.0295 -6.6069
#>   0.3721 -6.0039
#> 
#> (2,2,4,.,.) = 
#>   14.1826  -2.1120
#>   -2.4231  -1.1761
#>    3.9315   4.0719
#>    6.0762  -5.5093
#>    7.6248   5.0220
#> 
#> (3,2,4,.,.) = 
#>    2.9171  -6.1558
#>    0.6325   4.5176
#>   -4.8585  14.4394
#>   -2.1879  -0.9315
#>   -3.3200  -1.5535
#> 
#> (1,3,4,.,.) = 
#>   0.7813  7.7923
#>   0.2115  8.5660
#>  -10.1704  4.1674
#>   2.1576 -4.4216
#>  -1.7687 -1.7297
#> 
#> (2,3,4,.,.) = 
#>    7.2401   3.0097
#>   -3.5763   2.5399
#>    1.0425  -0.1574
#>   -3.4359  -1.4077
#>   10.6158  -2.1733
#> 
#> (3,3,4,.,.) = 
#>   9.3775 -11.8295
#>   0.9900  3.7263
#>  -1.8566 -4.9754
#>   3.1563  6.9626
#>   3.9723  3.1856
#> 
#> (1,1,5,.,.) = 
#>    3.4052  -3.3471
#>   10.3393  -8.3645
#>   -9.8148  -4.4791
#>   -1.3004   1.2262
#>    5.8919   2.7605
#> 
#> (2,1,5,.,.) = 
#>   4.3949  2.1508
#>  -7.0103  4.5990
#>   3.1547 -0.3337
#>  -4.5617 -4.2992
#>   4.8600 -2.1697
#> 
#> (3,1,5,.,.) = 
#>  -7.6660  4.6680
#>  -3.7334 -2.0026
#>  -1.5640 -7.2069
#>  -4.1978 -0.2492
#>   7.9738 -6.5265
#> 
#> (1,2,5,.,.) = 
#>  -2.8857 -1.5939
#>   1.7513  1.8288
#>   0.3072  1.1766
#>  -0.8078  1.8470
#>   4.4468  1.2259
#> 
#> (2,2,5,.,.) = 
#>   -5.3801   0.3765
#>   -2.3626  -2.7430
#>   -9.1551  -7.1093
#>    4.1213  11.0672
#>   -4.3100  -2.3626
#> 
#> (3,2,5,.,.) = 
#>  -5.5797  6.7338
#>  -12.8384  2.6177
#>   4.5008  0.8966
#>   2.6234  4.2010
#>   6.4968 -0.0604
#> 
#> (1,3,5,.,.) = 
#>   0.1716  0.7499
#>   7.0471  2.6079
#>  -1.4561  8.1277
#>  -3.7802  1.8000
#>  -0.0467 -0.5141
#> 
#> (2,3,5,.,.) = 
#>  -0.2791 -7.7801
#>   0.8033  9.8559
#>   0.1203  1.4078
#>   4.1299  0.2063
#>   9.5841  8.7177
#> 
#> (3,3,5,.,.) = 
#>  -0.7129 -8.8638
#>   0.7259  6.3403
#>   2.7860 -5.2647
#>  -3.0904 -0.9319
#>   1.7120 -5.2467
#> [ CPUFloatType{3,3,5,5,2} ]