Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   0.5351  2.9806
#>   1.8825 -7.5170
#>  -2.6263  1.4963
#>  -2.7044 -6.7158
#>  -8.5140 -3.6411
#> 
#> (2,1,1,.,.) = 
#>    1.4247  -0.7263
#>    2.0632  10.5222
#>   10.6883  -0.8873
#>    3.5544   0.2916
#>    1.4052  -2.2504
#> 
#> (3,1,1,.,.) = 
#>  -5.3301 -5.6192
#>   0.8712 -3.2836
#>  -0.5046  0.1313
#>   3.7635  1.0859
#>  -0.0703 -3.0767
#> 
#> (1,2,1,.,.) = 
#>    5.2453  -7.2805
#>   12.2623  -6.4644
#>    3.1998  -1.0653
#>    4.1277   3.8041
#>   -2.3649  -2.9722
#> 
#> (2,2,1,.,.) = 
#>  -1.3623  1.5839
#>   8.7085  5.3757
#>   3.1645 -5.8354
#>   1.1172 -8.2535
#>   3.0863  5.4030
#> 
#> (3,2,1,.,.) = 
#>   2.8111 -14.0949
#>  -4.4804 -5.4656
#>   3.8003  0.3617
#>   2.9558 -3.7827
#>  -0.8754 -3.0583
#> 
#> (1,3,1,.,.) = 
#>   -3.3666   4.8703
#>    2.5660  10.5149
#>   -4.3487  11.4440
#>    1.3896  -1.2429
#>   -6.1230  -3.5709
#> 
#> (2,3,1,.,.) = 
#>   6.8273 -1.2028
#>  -2.3022  4.9440
#>   3.5515  1.9656
#>   3.8174 -4.1475
#>  -0.0298  4.3850
#> 
#> (3,3,1,.,.) = 
#>   7.3511 -2.0283
#>   0.6425 -2.0348
#>  -9.8170 -10.0263
#>  -3.8116 -2.3452
#>   9.3357 -2.3784
#> 
#> (1,1,2,.,.) = 
#>   2.5919  1.7626
#>  -6.0684  7.3186
#>  -0.0384 -2.2817
#>   3.9593 -13.7005
#>   0.1256 -3.3274
#> 
#> (2,1,2,.,.) = 
#>  -0.6199  9.4028
#>   9.7861 -3.8426
#>   1.9064 -7.1686
#>   1.2816 -1.9541
#>   5.9310  7.2101
#> 
#> (3,1,2,.,.) = 
#>  -0.0780  5.7819
#>   5.2347  1.7154
#>   7.3643  0.2099
#>  -1.0237 -1.1604
#>   1.6152 -2.8934
#> 
#> (1,2,2,.,.) = 
#>   2.0193 -2.2434
#>   5.0929 -5.9610
#>  -5.6087  4.3393
#>   3.5354  1.7680
#>  -2.6485 -4.0102
#> 
#> (2,2,2,.,.) = 
#>   2.7529  0.9856
#>   0.6949  2.0424
#>   5.0894 -0.5660
#>   0.4392  0.6907
#>  -6.1236 -0.6993
#> 
#> (3,2,2,.,.) = 
#>  -2.5272 -0.2110
#>  -1.2424  6.7730
#>   9.1499  6.5191
#>   0.1720 -1.1697
#>   1.4853  1.9035
#> 
#> (1,3,2,.,.) = 
#>   -0.2498  -3.6154
#>    1.1320  -2.7438
#>    3.3655  -2.4004
#>    1.7099  10.5081
#>    2.1693  -1.8261
#> 
#> (2,3,2,.,.) = 
#>  -0.2612  4.8592
#>   0.1787 -9.2947
#>  -6.2396 -5.5702
#>  -4.5792 -3.9837
#>  -2.7556 -2.8758
#> 
#> (3,3,2,.,.) = 
#>  -11.7095   6.6501
#>   -5.9135   5.9642
#>    0.3424   0.4723
#>    1.8491   9.0261
#>    1.3820  -6.7759
#> 
#> (1,1,3,.,.) = 
#>  -4.4815 -2.5350
#>   2.7415  6.0914
#>  -15.3384  0.7293
#>  -3.7887 -0.8786
#>  -0.8085 -8.7210
#> 
#> (2,1,3,.,.) = 
#>   -4.7030   2.2650
#>   -6.8135   0.6237
#>   -4.0002  -2.8572
#>   -1.1293  15.3000
#>   -1.4613   2.5327
#> 
#> (3,1,3,.,.) = 
#>   4.7299  9.1314
#>  -3.2609 -3.2255
#>  -3.7850  0.5864
#>  -6.4529  5.6882
#>   8.3379  2.0731
#> 
#> (1,2,3,.,.) = 
#>  -1.2636  1.6989
#>  -10.7993  0.9868
#>   1.7147  5.6092
#>   2.0316  2.3903
#>   3.6770  6.7834
#> 
#> (2,2,3,.,.) = 
#>  -3.2150  2.1993
#>   0.7637  4.0776
#>   1.2905 -10.2239
#>   0.8787 -6.8202
#>  -0.6538 -0.7250
#> 
#> (3,2,3,.,.) = 
#>  -0.9179  0.6593
#>   1.0853 -0.6352
#>  -7.3530 -6.9246
#>   1.0629  7.9957
#>  -3.9713  2.0358
#> 
#> (1,3,3,.,.) = 
#>  -1.2120 -10.3264
#>   3.0497  9.1703
#>   6.3240  0.1729
#>   1.4546 -2.9780
#>   2.2779 -1.3136
#> 
#> (2,3,3,.,.) = 
#>    0.0707   3.2434
#>   -5.1021  -3.2082
#>    3.7848   7.8449
#>   -0.2646  -1.5225
#>    0.2617  10.5336
#> 
#> (3,3,3,.,.) = 
#>   -0.7220   4.0688
#>    0.1101  -8.7296
#>   11.0361   2.1472
#>    3.5283  -5.5593
#>    4.7233   3.2473
#> 
#> (1,1,4,.,.) = 
#>   4.0997 -3.1923
#>   6.3084 -6.1616
#>  -0.6268  0.7301
#>   9.9664  3.3424
#>  -4.7224 -2.1575
#> 
#> (2,1,4,.,.) = 
#>   -2.7975  -7.2717
#>    5.2979  11.8373
#>   -2.8179  -0.0141
#>   -2.8905   5.3211
#>    2.2221   0.2257
#> 
#> (3,1,4,.,.) = 
#>  -3.3918  0.0680
#>   0.5325  5.8076
#>   0.6076  9.5723
#>  -0.9397 -2.9039
#>   2.4073  1.5955
#> 
#> (1,2,4,.,.) = 
#>   4.4767 -7.2413
#>  -8.8671 -3.8322
#>   3.9996 -0.1496
#>  -0.4243  5.0160
#>   3.0946 -3.2432
#> 
#> (2,2,4,.,.) = 
#>   4.2251  2.1978
#>  -3.8433  2.2746
#>  -1.2726  4.7449
#>  -8.5752 -1.3999
#>   4.0684  5.9567
#> 
#> (3,2,4,.,.) = 
#>  -5.3562  1.2706
#>   5.7350  5.3445
#>   3.2276  1.1356
#>   0.5956 -0.2391
#>  -3.5493  1.8516
#> 
#> (1,3,4,.,.) = 
#>   -5.6885  -2.9883
#>    5.4686  11.0354
#>   -3.1585  -5.5320
#>   -2.5147   0.4970
#>   -4.7337 -10.0027
#> 
#> (2,3,4,.,.) = 
#>  -4.2544  3.5175
#>   1.7121 -1.3197
#>  -4.4301  2.7117
#>   2.2718  5.2550
#>  -8.0238  5.5961
#> 
#> (3,3,4,.,.) = 
#>   2.6784  1.5256
#>   3.0406  1.8892
#>  -6.7574 -0.2310
#>  -1.5660  4.2380
#>  -8.4748 -4.4146
#> 
#> (1,1,5,.,.) = 
#>  -0.4063  1.4347
#>   4.3156 -2.5606
#>  -8.5631  1.3290
#>   5.2914 -1.8464
#>  -0.5408  4.3965
#> 
#> (2,1,5,.,.) = 
#>   2.1060  1.4650
#>   5.8816  8.7728
#>  -5.4176  5.4857
#>  -3.1968  0.6906
#>   2.6729  3.7597
#> 
#> (3,1,5,.,.) = 
#>  -1.2821 -1.7923
#>  -2.3624  0.5954
#>  -6.7107 -1.4803
#>   4.0450 -3.4223
#>  -5.3375  1.7821
#> 
#> (1,2,5,.,.) = 
#>  -2.3276 -2.9040
#>   0.9719 -3.3857
#>  -4.4820  1.7338
#>   0.5525 -1.6901
#>  -12.6577 -3.6604
#> 
#> (2,2,5,.,.) = 
#>  -4.4096  3.7270
#>   0.3239 -0.8539
#>   7.4867 -4.5046
#>   7.9285 -2.2470
#>   1.5404  5.5138
#> 
#> (3,2,5,.,.) = 
#>   -1.1540   2.9924
#>   -5.0539  -4.2369
#>   -3.6304  -4.8502
#>   -0.8053   4.0432
#>    2.2966  10.8617
#> 
#> (1,3,5,.,.) = 
#>    7.8599  -1.4364
#>   -1.2374  -4.1721
#>   -2.9784  11.8586
#>    5.3416   1.5225
#>  -10.0656   0.9673
#> 
#> (2,3,5,.,.) = 
#>   -3.7150  -1.3540
#>   13.8904   1.2721
#>    6.2505  -2.5721
#>   -3.1341   3.4308
#>    8.9375 -10.3441
#> 
#> (3,3,5,.,.) = 
#>  -3.4631  2.9072
#>  -6.0030  0.4438
#>  -4.2079  0.9144
#>  -3.9245 -1.6894
#>   1.6268 -4.0685
#> [ CPUFloatType{3,3,5,5,2} ]