Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   1.8299 -0.8545
#>   4.8936  5.5387
#>   8.1568 -5.3317
#>  -8.0713 -0.4569
#>   2.3043 -1.2616
#> 
#> (2,1,1,.,.) = 
#>   1.5061 -2.5082
#>  -7.2081  3.6199
#>   4.1284  8.2209
#>  -2.2337  3.8781
#>  -2.2533  0.9721
#> 
#> (3,1,1,.,.) = 
#>   5.4609 -10.1583
#>   7.4753  0.0960
#>  -5.1140 -1.1828
#>   5.8334  2.3750
#>   4.5559  0.0765
#> 
#> (1,2,1,.,.) = 
#>   5.6610  3.2235
#>   7.0432 -5.5988
#>  -10.3719  7.4866
#>  -10.5179  6.9639
#>   0.9828  2.1239
#> 
#> (2,2,1,.,.) = 
#>   10.5028  -1.5392
#>   -1.4293  -3.5677
#>   -5.2553  -0.2220
#>    1.0801   4.6798
#>   -2.8195  -5.4541
#> 
#> (3,2,1,.,.) = 
#>   2.7805  1.5709
#>   9.6489  0.4143
#>   3.7712  4.1986
#>  -10.5203  2.9443
#>   1.7063  4.4798
#> 
#> (1,3,1,.,.) = 
#>   -3.8016 -10.2178
#>   -3.3357   0.2338
#>    2.8052  11.0760
#>   -9.0456  -2.8879
#>    7.6091   4.3841
#> 
#> (2,3,1,.,.) = 
#>   -3.5916   5.0966
#>    3.6364  -1.0758
#>   -7.2250  -2.9791
#>   -1.0007  10.1377
#>    3.0983  -5.4333
#> 
#> (3,3,1,.,.) = 
#>  -4.4079  8.6195
#>   2.8221 -5.2594
#>  -4.8097 -2.5593
#>   7.0129  4.4923
#>  -5.1272 -4.0693
#> 
#> (1,1,2,.,.) = 
#>  -12.3206  -0.7139
#>   -1.2082   8.7472
#>   -8.8089   2.7068
#>   -0.3002  -4.1436
#>   -3.9514   5.3151
#> 
#> (2,1,2,.,.) = 
#>  -0.8914  4.8220
#>  -3.0093  0.3267
#>  -1.3298  4.6909
#>  -0.8826 -5.2531
#>  -0.5382 -5.9600
#> 
#> (3,1,2,.,.) = 
#>    2.8440  -4.8942
#>    1.2760  -1.3042
#>  -11.4849  -1.7285
#>    4.8920  10.6377
#>    1.3073  -0.8840
#> 
#> (1,2,2,.,.) = 
#>  -10.2702  -0.1947
#>    4.2283  -5.0279
#>    4.2727  14.2724
#>   -6.4569   6.7210
#>    6.7015   1.2868
#> 
#> (2,2,2,.,.) = 
#>   2.6739 -2.3141
#>   0.2269 -7.0764
#>  -1.5479  2.0028
#>   0.5153  0.8697
#>  -4.1571  3.8501
#> 
#> (3,2,2,.,.) = 
#>   2.2899 -6.4572
#>   0.6331  1.9728
#>  -2.8577 -5.8199
#>   0.3792 -0.2344
#>   9.7437  0.3884
#> 
#> (1,3,2,.,.) = 
#>   4.4693 -14.4161
#>   2.9678 -8.4648
#>   3.0824  5.1941
#>   5.9856 -8.4170
#>  -4.1002 -0.2242
#> 
#> (2,3,2,.,.) = 
#>   1.3738 -13.2057
#>   5.6119 -5.8169
#>   9.7263 -4.9929
#>  -2.0635 -3.2370
#>  -8.4518  3.9892
#> 
#> (3,3,2,.,.) = 
#>   -0.7289  -4.6840
#>    2.8690  -1.9408
#>    3.6092   6.2359
#>   -7.1584   3.3060
#>    2.7515  10.3662
#> 
#> (1,1,3,.,.) = 
#>  -9.0970  2.8348
#>   7.8277 -3.8625
#>  -4.5697  3.8870
#>   0.7759 -3.8838
#>   3.3643  4.3610
#> 
#> (2,1,3,.,.) = 
#>  -6.1251  0.0861
#>   3.2074 -3.1752
#>   8.1331 -1.5727
#>  -0.8345 -3.0852
#>  -2.6750 -6.5429
#> 
#> (3,1,3,.,.) = 
#>  -0.2310  0.4213
#>  -5.3056  2.6728
#>  -0.7398  0.9795
#>   1.8693  6.9272
#>  -11.7805  2.9388
#> 
#> (1,2,3,.,.) = 
#>   -2.0775   0.9143
#>    6.7132  -8.8991
#>    1.3144   2.2212
#>   -6.8120   0.7909
#>    9.6971  11.4349
#> 
#> (2,2,3,.,.) = 
#>  -9.8568 -8.5726
#>   7.6352 -8.9431
#>  -4.7963 -6.1808
#>   0.8898  3.0809
#>   1.1815  5.2101
#> 
#> (3,2,3,.,.) = 
#>   8.6539 -7.6683
#>  -0.9707 -0.0203
#>   4.4724  0.0112
#>  -2.9278 -1.8548
#>   0.7622  0.2085
#> 
#> (1,3,3,.,.) = 
#>    0.6177  -3.9783
#>    9.1741   1.3555
#>   11.1847   4.0785
#>    2.7688   1.8064
#>   -0.1331  -0.5537
#> 
#> (2,3,3,.,.) = 
#>  -0.7404 -3.8229
#>   1.3602  1.3962
#>  -4.0166 -10.0473
#>   0.7993  0.6364
#>   1.4918  2.8617
#> 
#> (3,3,3,.,.) = 
#>  -6.6438 -5.8591
#>  -10.0200  2.2828
#>  -6.3160  3.5947
#>  -7.9746 -2.4736
#>  -0.6530 -2.8422
#> 
#> (1,1,4,.,.) = 
#>   3.3780  6.1701
#>   2.7823 -1.7109
#>   1.5653  4.1627
#>  -0.9770  3.4156
#>  -3.1627  7.6722
#> 
#> (2,1,4,.,.) = 
#>   -4.2459  -3.3951
#>   -0.2220 -10.5579
#>   -3.2891   0.1922
#>    4.9543  13.6104
#>   -5.6967   0.5392
#> 
#> (3,1,4,.,.) = 
#>  -3.5985 -7.0894
#>   2.5169 -1.2003
#>   3.5668  0.6172
#>  -2.9016  7.9270
#>   0.2126 -2.4786
#> 
#> (1,2,4,.,.) = 
#>  -0.7997  1.6436
#>  -9.0900  8.1134
#>  -1.5759  7.8436
#>   6.2540 -10.4366
#>  -5.2340  6.7355
#> 
#> (2,2,4,.,.) = 
#>   3.8949  8.6516
#>  -4.8663  2.2308
#>  -0.2335  5.6199
#>   4.2354 -4.1996
#>   7.0605 -0.8072
#> 
#> (3,2,4,.,.) = 
#>  -3.5577  1.4423
#>   3.6434 -1.7864
#>   1.6322  8.4587
#>  -2.6542 -4.8610
#>  -0.8034  7.5955
#> 
#> (1,3,4,.,.) = 
#>   2.3714 -6.9130
#>   8.3922 -2.1175
#>  -1.5761 -7.0675
#>   3.9819 -5.7787
#>   8.8816  2.8237
#> 
#> (2,3,4,.,.) = 
#>   5.5048 -2.6695
#>   1.9431 -1.9694
#>   1.2273  3.8493
#>   0.0113 -0.6937
#>   1.6932 -2.2065
#> 
#> (3,3,4,.,.) = 
#>   -3.6905   5.8336
#>   -0.6594   1.7002
#>    5.8126 -16.8538
#>    6.2780   5.2017
#>   11.0860   1.8238
#> 
#> (1,1,5,.,.) = 
#>  -3.3967 -3.2141
#>   0.1716  5.0541
#>  -1.1075 -6.5374
#>   1.8133 -2.1525
#>   5.1981  1.9159
#> 
#> (2,1,5,.,.) = 
#>  -6.2180  0.0968
#>  -5.9405 -2.8848
#>   9.8362 -7.0165
#>   2.8734 -2.3316
#>  -0.0163  0.5031
#> 
#> (3,1,5,.,.) = 
#>  -3.8829  1.1589
#>  -4.3437  7.6206
#>  -0.9220 -10.3968
#>   0.2424  3.8622
#>   0.9469 -9.6853
#> 
#> (1,2,5,.,.) = 
#>   9.1159 -3.3963
#>   0.9571  3.4686
#>  -6.5812 -1.1398
#>   2.3164  0.8760
#>  -9.0979  1.2900
#> 
#> (2,2,5,.,.) = 
#>    1.9017  -1.0171
#>    8.7321   3.4119
#>   -2.9992  -5.4508
#>   -6.0562  -5.0832
#>   -1.0171  10.9275
#> 
#> (3,2,5,.,.) = 
#>   0.2695 -2.9452
#>   5.8295 -0.9170
#>  -1.7834 -0.7470
#>  -10.5062  7.9936
#>   0.7503 -0.0258
#> 
#> (1,3,5,.,.) = 
#>   4.6380 -8.7288
#>  -4.0617  3.8543
#>  -10.7917 -2.2960
#>  -2.3112 -1.9079
#>  -1.6351 -2.0068
#> 
#> (2,3,5,.,.) = 
#>    2.7100   1.8331
#>    0.9524  -3.4304
#>    1.6744  -8.1538
#>   13.8007 -15.6733
#>   -6.5618   5.0173
#> 
#> (3,3,5,.,.) = 
#>   0.7307 -5.4273
#>   3.8142  4.0542
#>  -2.6337  3.7278
#>   6.2772  6.9084
#>  -0.9076 -8.7595
#> [ CPUFloatType{3,3,5,5,2} ]