Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   9.3401 -7.3237
#>   6.4594  9.1496
#>  -0.3859 -5.0341
#>  -6.7796 -3.3438
#>   1.1873 -0.8582
#> 
#> (2,1,1,.,.) = 
#>  -2.5846  4.1842
#>  -3.9746 -4.3075
#>   0.4610  2.6198
#>   2.3267 -9.2056
#>   3.5879 -8.4836
#> 
#> (3,1,1,.,.) = 
#>   5.3441 -5.0803
#>   1.1280  1.5648
#>   5.8840  4.2929
#>   0.2423 -5.6848
#>   3.3878  3.5818
#> 
#> (1,2,1,.,.) = 
#>  -6.3136 -7.0616
#>  -2.7010 -9.6344
#>  -6.4962  0.9699
#>  -8.4824  8.7951
#>  -2.2852 -6.1216
#> 
#> (2,2,1,.,.) = 
#>  -3.4945  1.4583
#>   6.7004 -3.2429
#>   3.8885 -3.1058
#>   2.5254 -8.0346
#>  -0.1249  5.5948
#> 
#> (3,2,1,.,.) = 
#>  -3.7537  2.2916
#>   4.3000  5.5464
#>  -2.0370  0.4952
#>   8.2523  1.3932
#>  -0.2216 -1.1073
#> 
#> (1,3,1,.,.) = 
#>   0.1399 -8.0163
#>  -1.2463  9.7061
#>  -1.2899  9.9311
#>   3.3586  1.4495
#>  -4.2439  1.5648
#> 
#> (2,3,1,.,.) = 
#>  -1.6031 -10.1580
#>  -1.2546  2.6541
#>  -2.7339 -0.5146
#>   4.3064  2.3303
#>   4.3952 -2.2229
#> 
#> (3,3,1,.,.) = 
#>   -2.9900   2.0936
#>   -3.8181   3.1574
#>    6.9513  -1.3982
#>   -2.6682  10.2070
#>    4.8906   4.6482
#> 
#> (1,1,2,.,.) = 
#>  -3.8340 -0.6995
#>   2.0827 -3.2400
#>  -1.1898  7.0873
#>  -1.1537  2.9571
#>   8.6426  9.1985
#> 
#> (2,1,2,.,.) = 
#>   -3.8821   0.5305
#>    2.6432   3.3540
#>   -1.3594   6.4822
#>   15.4160   4.6630
#>   -6.7182  -0.7639
#> 
#> (3,1,2,.,.) = 
#>    4.1650   2.9880
#>  -13.7410 -10.0129
#>    2.3187  11.1032
#>   -4.0209  -5.7597
#>   -3.9356   6.2103
#> 
#> (1,2,2,.,.) = 
#>  -10.2607   8.5069
#>   -3.8332   1.4950
#>    3.4100  -7.6663
#>   -6.1280   0.9663
#>    7.1347  -4.9448
#> 
#> (2,2,2,.,.) = 
#>   2.8940  4.8416
#>   6.4610  0.4308
#>   1.9927 -6.3210
#>   8.0570 -2.5799
#>  -1.9457  2.1048
#> 
#> (3,2,2,.,.) = 
#>   4.4560 -4.3274
#>  -3.7387  1.9864
#>   5.9856  4.8884
#>  -3.7835 -2.1303
#>   5.7988 -0.1539
#> 
#> (1,3,2,.,.) = 
#>  -3.4122  6.1820
#>  -2.8696 -6.1495
#>   0.5221 -4.7845
#>  -8.9451  0.9784
#>   1.9214  4.5823
#> 
#> (2,3,2,.,.) = 
#>  -1.5985 -4.3374
#>  -4.8008 -5.7859
#>  -4.1257  4.3387
#>  -4.6931 -2.4641
#>  -4.3318  5.2979
#> 
#> (3,3,2,.,.) = 
#>  -1.0220  4.3723
#>  -3.4789 -5.3568
#>  -4.7613  6.6014
#>  -3.9829 -3.5713
#>   0.9223  6.1366
#> 
#> (1,1,3,.,.) = 
#>   -9.2202  -3.6041
#>   -2.6618  15.6429
#>    8.2854   1.0139
#>    3.1328  -4.1691
#>   -0.0681  12.8891
#> 
#> (2,1,3,.,.) = 
#>   1.6376  0.7217
#>   1.9824  4.1926
#>   0.5200  8.4328
#>   5.4769  5.6431
#>  -7.0022  4.6521
#> 
#> (3,1,3,.,.) = 
#>  -6.3985 -5.6721
#>   0.7751 -2.6204
#>  -9.0504 -0.0023
#>  -5.8121 -1.0367
#>  -10.7140 -8.9181
#> 
#> (1,2,3,.,.) = 
#>    2.3898   1.2634
#>   -6.9024   6.8861
#>    5.7858  14.2334
#>   -0.5018   9.1893
#>    6.2739  -2.8583
#> 
#> (2,2,3,.,.) = 
#>  -1.8841 -3.0618
#>  -3.3355 -5.3030
#>  -4.7974  1.3572
#>   1.3757  5.3942
#>  -3.5328 -0.9889
#> 
#> (3,2,3,.,.) = 
#>    5.3877  -6.9668
#>    2.5698   2.8214
#>    8.3492  -0.8881
#>   13.1933   6.6061
#>   -8.9367   5.4356
#> 
#> (1,3,3,.,.) = 
#>   1.8499  9.1275
#>  -3.5549  4.4921
#>  -1.2433 -7.3614
#>  -0.0277  0.3861
#>  -1.6492  2.4603
#> 
#> (2,3,3,.,.) = 
#>   -3.9931  -0.9929
#>   -1.7030  -6.8069
#>   11.6603  -0.2560
#>   -4.1613   4.2966
#>    2.5387   3.1961
#> 
#> (3,3,3,.,.) = 
#>  -1.3930 -3.7577
#>  -4.9575 -2.8934
#>  -4.5920  8.6475
#>   7.1622  1.0966
#>  -7.3040 -6.5927
#> 
#> (1,1,4,.,.) = 
#>  -6.0001  2.6596
#>   3.5305  3.1406
#>  -4.4788 -0.9823
#>   2.0550 -3.8295
#>   5.0558 -1.6409
#> 
#> (2,1,4,.,.) = 
#>  -1.1844  7.5260
#>  -1.1469  0.5581
#>  -1.8704 -2.9036
#>   4.4378 -0.2389
#>   2.1649  3.3892
#> 
#> (3,1,4,.,.) = 
#>   -3.3427   0.7437
#>    4.2762   3.8197
#>   -2.0549  -2.6173
#>    4.0930   3.2048
#>   10.0333   6.7624
#> 
#> (1,2,4,.,.) = 
#>   1.7411 -3.5196
#>  -0.7263 -2.9802
#>   8.6303  0.5665
#>   1.3218  4.6741
#>   5.9841  2.3032
#> 
#> (2,2,4,.,.) = 
#>   6.1746  0.4210
#>  -0.2207  3.0045
#>  -3.1838  1.4591
#>  -9.8779  5.9137
#>   2.4743 -2.7609
#> 
#> (3,2,4,.,.) = 
#>    4.9635  10.9868
#>    2.8711  -9.1639
#>   -7.1545  -0.9152
#>   -6.6691  -1.3626
#>   -9.2242   2.1931
#> 
#> (1,3,4,.,.) = 
#>  -4.5591 -2.3398
#>  -6.3338 -9.3517
#>   1.9459 -7.5547
#>  -2.0534  3.3609
#>   4.7304 -3.9643
#> 
#> (2,3,4,.,.) = 
#>  -5.7755  2.8506
#>   2.9856 -6.4090
#>  -2.2843  4.1464
#>  -7.9151 -7.4412
#>  -4.2086  5.6825
#> 
#> (3,3,4,.,.) = 
#>   -1.4017  -1.9138
#>   -2.4577   9.1968
#>    1.8257   4.6245
#>   -8.9736  -6.5392
#>   10.3821   0.7261
#> 
#> (1,1,5,.,.) = 
#>  -2.6433 -4.8273
#>   5.2748 -1.7604
#>  -0.7452 -4.1837
#>   1.5853 -1.1057
#>  -8.6237  0.6659
#> 
#> (2,1,5,.,.) = 
#>  -0.6438  0.8730
#>   1.6796  2.4241
#>  -8.8617 -3.8732
#>   0.6511  3.2638
#>  -2.4276  0.0387
#> 
#> (3,1,5,.,.) = 
#>   3.0157 -6.5440
#>   1.1595  6.7273
#>  -6.3753 -1.7177
#>  -5.1012  8.9662
#>  -2.1171  5.0481
#> 
#> (1,2,5,.,.) = 
#>   2.7694  6.6697
#>  -1.2788 -1.9138
#>  -6.4888 -9.5106
#>   0.2171 -7.4289
#>   1.4208 -6.7625
#> 
#> (2,2,5,.,.) = 
#>   -0.4476  11.5453
#>   -5.3879   5.4658
#>    0.9246   6.1739
#>    3.9347 -13.0866
#>    9.0060   0.5022
#> 
#> (3,2,5,.,.) = 
#>   5.7223  6.6437
#>  -11.0499 -0.6866
#>   2.1010  2.7412
#>  -2.6058  2.3912
#>  -6.2422  7.8957
#> 
#> (1,3,5,.,.) = 
#>   -0.3468   2.2732
#>    1.5684  13.7393
#>   -8.8475   6.5673
#>   -3.6624   0.2302
#>   -0.6703  -7.9039
#> 
#> (2,3,5,.,.) = 
#>  -3.3014  0.4259
#>   2.6499  2.4296
#>  -0.3102 -7.0941
#>   0.2796  6.0119
#>   1.5694  1.0166
#> 
#> (3,3,5,.,.) = 
#>   -9.9383  -0.1358
#>   -6.9880  -3.3504
#>    3.1010  -5.4246
#>   -4.5843  11.1072
#>   -7.0850  -0.9335
#> [ CPUFloatType{3,3,5,5,2} ]