Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -7.0064 -3.6828
#>  -2.8180 -9.1078
#>   0.6607  0.8451
#>  -3.2385 -3.1249
#>   8.0422 -0.4805
#> 
#> (2,1,1,.,.) = 
#>  -0.5100 -5.1461
#>  -3.1645 -1.7265
#>   5.2166  6.2279
#>  -3.4418 -0.9923
#>   1.9166 -0.6176
#> 
#> (3,1,1,.,.) = 
#>   1.7067 -3.2049
#>  -1.2279  3.6395
#>  -0.6172 -5.9008
#>  -0.0824 -7.3495
#>   1.9054  6.9163
#> 
#> (1,2,1,.,.) = 
#>  -1.3624  9.1987
#>   4.8035  0.3220
#>   1.4580 -4.5193
#>   1.9154 -0.8220
#>  -6.1195  1.7628
#> 
#> (2,2,1,.,.) = 
#>  -7.0654  1.7418
#>  -5.4333 -1.5132
#>  -0.7208  6.2347
#>   7.4351 -1.1753
#>   5.1756 -0.5529
#> 
#> (3,2,1,.,.) = 
#>   -4.0025 -12.1808
#>    0.0017   7.1278
#>   14.4784   0.0832
#>   -0.6058  -3.0031
#>   -3.6196   4.8929
#> 
#> (1,3,1,.,.) = 
#>  -0.2095 -1.5311
#>   3.0555  1.6991
#>  -1.1506  2.2252
#>  -3.3322 -1.1813
#>  -1.8608 -1.8267
#> 
#> (2,3,1,.,.) = 
#>   -2.7655  -2.7502
#>   -1.9583  -3.2407
#>    2.8734  10.2167
#>   -1.3410   1.1103
#>    6.7157   3.7633
#> 
#> (3,3,1,.,.) = 
#>    3.5598   1.2367
#>    4.8571   1.9015
#>    0.8840   8.2691
#>   -2.6996  11.5990
#>   -2.4286   2.3230
#> 
#> (1,1,2,.,.) = 
#>  -8.0071  4.8570
#>   0.9814 -7.2368
#>   1.3475  6.5414
#>   1.5948 -4.0251
#>  -1.5003  3.4454
#> 
#> (2,1,2,.,.) = 
#>   0.3720 -7.5198
#>   1.5772  2.3606
#>   2.9556 -8.0018
#>  -0.8281 -4.1876
#>   3.3460  4.1269
#> 
#> (3,1,2,.,.) = 
#>   1.3417  8.2708
#>  -2.2240  8.0031
#>   3.9355 -0.1998
#>  -3.7462  7.7785
#>   2.3917  4.2830
#> 
#> (1,2,2,.,.) = 
#>  -12.2818   4.1862
#>   -4.9662  -2.2931
#>   -2.8482   5.0200
#>   -1.2973   0.0971
#>   -7.4537  -0.7719
#> 
#> (2,2,2,.,.) = 
#>   -3.1597   0.0487
#>   -0.0924   2.7802
#>    6.7851  -8.8806
#>    0.0626   3.7290
#>   10.3873  -5.2515
#> 
#> (3,2,2,.,.) = 
#>  -0.5866 -1.6328
#>   1.0262 -2.3877
#>  -3.2980 -1.2526
#>   1.3675 -2.6380
#>   0.6851  2.6325
#> 
#> (1,3,2,.,.) = 
#>  -1.7367 -6.5280
#>  -6.3431  3.0075
#>   6.2638 -2.6870
#>  -13.4444 -1.4753
#>  -6.5971 -10.8609
#> 
#> (2,3,2,.,.) = 
#>  -4.2452  3.9928
#>   6.1524 -0.8988
#>   2.0676 -2.5549
#>   4.1435 -5.3995
#>  -0.3471  2.6605
#> 
#> (3,3,2,.,.) = 
#>  -1.7633  6.4231
#>  -3.4377 -2.7773
#>   0.5232  4.7301
#>  -5.4203  0.6097
#>   2.9244 -5.8965
#> 
#> (1,1,3,.,.) = 
#>  -4.1555 -2.2894
#>   8.7500  2.5655
#>   2.6199  1.2056
#>   5.3022  0.9489
#>  -1.2319 -4.5115
#> 
#> (2,1,3,.,.) = 
#>   7.5317 -0.9210
#>  -0.8603  0.1131
#>   6.7314  5.8708
#>  -3.5068  2.2437
#>  -3.7502 -3.4403
#> 
#> (3,1,3,.,.) = 
#>    1.4007   0.8399
#>    0.2278  -4.3782
#>  -10.7868   7.5771
#>    6.6498  12.7030
#>   -4.1548  -1.9282
#> 
#> (1,2,3,.,.) = 
#>   2.6268 -0.2360
#>  -2.9532 -0.7405
#>   1.2660  5.8508
#>   5.6726 -1.6416
#>   2.1833  2.0676
#> 
#> (2,2,3,.,.) = 
#>  -8.0924  2.0213
#>  -2.1805  0.8303
#>  -0.5348  0.6469
#>  -6.3481  2.2381
#>  -1.3417  5.6679
#> 
#> (3,2,3,.,.) = 
#>   11.3971  -5.4738
#>   -1.4021  -4.6691
#>    5.7172  -4.6449
#>    5.8261  -2.2775
#>   -0.5459   0.6143
#> 
#> (1,3,3,.,.) = 
#>  -1.3700 -0.8252
#>   4.5053 -1.9979
#>   0.4034 -3.6250
#>  -4.6099 -1.7676
#>  -2.1800 -0.5169
#> 
#> (2,3,3,.,.) = 
#>   9.8984 -1.6404
#>   2.2706  8.4176
#>   1.7645 -5.9546
#>  -0.3268  5.7981
#>   1.3437 -5.1399
#> 
#> (3,3,3,.,.) = 
#>    1.3610  -2.7140
#>    1.5249   1.1454
#>   -4.7623   0.7162
#>   -3.3823  -0.4585
#>    7.0338  13.0753
#> 
#> (1,1,4,.,.) = 
#>   2.7652 -1.9079
#>  -6.1002  1.5280
#>   0.6581 -2.4523
#>  -5.5581  7.6510
#>  -3.9469  2.3982
#> 
#> (2,1,4,.,.) = 
#>   5.8283 -10.5574
#>  -0.3151  5.8643
#>   2.1811  0.8711
#>   9.0367  2.2959
#>  -0.5017  0.3260
#> 
#> (3,1,4,.,.) = 
#>    1.0336  -8.6536
#>   -4.1474  -2.4881
#>   -4.0150   0.5436
#>    0.2604  -1.7849
#>   11.9293  -4.3028
#> 
#> (1,2,4,.,.) = 
#>   1.7425 -2.2194
#>   4.0134  1.9141
#>   1.5953 -0.5478
#>  -0.5487 -0.9328
#>  -3.1269 -4.6500
#> 
#> (2,2,4,.,.) = 
#>   11.5231  -5.8390
#>   -4.4155  -5.2269
#>   -5.4092 -10.4901
#>    3.9829  -3.0336
#>    0.2022  -0.9372
#> 
#> (3,2,4,.,.) = 
#>  -7.7858 -3.4562
#>  -8.2040 -11.4615
#>  -0.3684  1.5672
#>  -2.6922  2.1000
#>   0.2121 -3.5731
#> 
#> (1,3,4,.,.) = 
#>   0.8112  2.5676
#>  -11.2660  5.0631
#>  -5.7830  7.0058
#>   5.9321  5.2540
#>   1.5578  4.5085
#> 
#> (2,3,4,.,.) = 
#>   3.9560 -7.1638
#>  -14.8701  6.0426
#>   4.9331 -0.9652
#>   3.2147 -2.9251
#>   1.3546 -5.5793
#> 
#> (3,3,4,.,.) = 
#>  -0.1443 -0.6667
#>   1.5054 -0.4761
#>   1.2540  5.0159
#>   3.2131  8.0633
#>  -1.0626  6.2703
#> 
#> (1,1,5,.,.) = 
#>   6.1247  6.1206
#>  -6.0366 -4.1997
#>  -0.0104  6.7615
#>   2.5485  6.1190
#>   3.3782  0.2780
#> 
#> (2,1,5,.,.) = 
#>   4.2564 -2.6147
#>   0.2033 -3.2209
#>   5.7148 -0.8566
#>  -2.6180 -1.0646
#>   7.2606  2.5314
#> 
#> (3,1,5,.,.) = 
#>   2.6996  5.7667
#>   0.6289 -2.1067
#>   6.1028  1.5725
#>   3.7233  1.4296
#>  -5.1664  8.9568
#> 
#> (1,2,5,.,.) = 
#>  -0.4881 -3.6272
#>  -2.2379 -0.7219
#>  -3.3345 -0.2720
#>   0.8352  2.4750
#>  -2.6596 -5.2554
#> 
#> (2,2,5,.,.) = 
#>   1.9872  6.2382
#>   4.3455  3.0275
#>  -3.0201 -2.9803
#>   2.4596  3.4754
#>   6.3150 -0.3716
#> 
#> (3,2,5,.,.) = 
#>    3.7536  10.5467
#>    3.8397   0.0088
#>    7.1402   0.3846
#>   -0.6239  -2.7950
#>    1.0393   2.8896
#> 
#> (1,3,5,.,.) = 
#>    4.4123  11.4413
#>    2.2069  -0.0949
#>    4.6406   4.5408
#>  -10.1232   0.7568
#>   -3.7930  -3.0814
#> 
#> (2,3,5,.,.) = 
#>   0.7990 -1.8233
#>   0.7197 -5.7273
#>  -3.9614 -0.6741
#>   4.0811 -0.7645
#>   1.4376  1.7461
#> 
#> (3,3,5,.,.) = 
#>  -3.5325 -1.2894
#>   3.1636 -1.7475
#>  -4.5833  5.5333
#>   1.5381 -1.5060
#>   2.5105  1.3820
#> [ CPUFloatType{3,3,5,5,2} ]