Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   1.9763  7.8467
#>  -4.3943 -2.8071
#>   8.3070  4.7473
#>   0.1214  9.9011
#>  -0.4568  1.8810
#> 
#> (2,1,1,.,.) = 
#>    1.4268  -3.2046
#>    8.7205   1.2440
#>   -6.5158  -2.4400
#>    3.4382  11.7033
#>    1.6350   2.4011
#> 
#> (3,1,1,.,.) = 
#>   6.0755  3.5094
#>   4.9627  8.6584
#>   4.3683 -0.5855
#>  -1.6353 -0.9633
#>   6.7713  6.0062
#> 
#> (1,2,1,.,.) = 
#>  -7.9359 -5.3996
#>  -10.9992  1.7149
#>   4.6450  0.8580
#>  -0.9310 -0.2429
#>   3.7647  6.5126
#> 
#> (2,2,1,.,.) = 
#>  -11.2618  -2.7025
#>    7.7288   2.3066
#>    4.5014   1.8962
#>    7.0432  -8.6552
#>   -1.4062  -3.4954
#> 
#> (3,2,1,.,.) = 
#>  -2.3604 -3.6922
#>  -0.1313  2.1074
#>   3.2983  3.5613
#>  -2.5760  2.7359
#>   0.8110 -7.0844
#> 
#> (1,3,1,.,.) = 
#>  -8.4593 -0.1551
#>   0.5104 -6.3483
#>   6.8818 -11.5768
#>  -3.7438  3.1692
#>   0.4190 -5.1557
#> 
#> (2,3,1,.,.) = 
#>  -1.1458 -11.0048
#>  -0.9352  5.5999
#>   3.9602 -5.6657
#>   8.5348 -3.0511
#>   3.4337  1.3237
#> 
#> (3,3,1,.,.) = 
#>  -3.0854 -2.5195
#>  -3.3872  0.8570
#>   2.3626  3.4360
#>  -0.9504  1.7517
#>  -4.4356 -8.5422
#> 
#> (1,1,2,.,.) = 
#>  -1.6043  4.3357
#>   1.7267 -2.7912
#>  -3.2619 -0.0287
#>  -3.7177  3.7267
#>  -6.2791  4.5578
#> 
#> (2,1,2,.,.) = 
#>  -4.7756 -5.8334
#>  -8.0739 -1.0256
#>  -6.2643  4.3290
#>  -3.6430 -1.2482
#>  -6.7862  6.6639
#> 
#> (3,1,2,.,.) = 
#>  -5.2635 -2.6365
#>   0.8090 -10.0458
#>  -4.8289 -4.6526
#>  -1.3182 -3.2208
#>  -3.6673  1.9778
#> 
#> (1,2,2,.,.) = 
#>  -10.5196   7.4601
#>    3.2538   3.9547
#>   -5.7687  -3.8533
#>    1.9987  -1.9309
#>   -1.5664   0.3644
#> 
#> (2,2,2,.,.) = 
#>  -13.2150  -9.4059
#>    0.3705  -1.1361
#>    0.3318   4.9082
#>    3.4515   5.6696
#>    2.4051  -7.6619
#> 
#> (3,2,2,.,.) = 
#>  -5.0949 -0.7683
#>  -6.1067 -2.4366
#>   1.2538  0.3982
#>  -2.7778  1.5682
#>  -1.2664  2.9718
#> 
#> (1,3,2,.,.) = 
#>  -3.4493  3.2682
#>   1.0988 -3.9018
#>  -1.9195  2.6591
#>  -0.0942 -1.2211
#>  -3.2207  3.5987
#> 
#> (2,3,2,.,.) = 
#>   -2.6009  -4.5345
#>   -0.9203   6.5752
#>    2.2495  -4.3585
#>   10.6067   1.5344
#>    0.8452   0.4637
#> 
#> (3,3,2,.,.) = 
#>   4.2682 -3.6109
#>   4.0290  3.0624
#>   0.5968 -2.8749
#>  -0.9304 -0.8756
#>  -4.1537  2.7282
#> 
#> (1,1,3,.,.) = 
#>   10.4268   0.2844
#>    3.0158   5.1906
#>   -5.6678  11.2523
#>   10.2783   0.1301
#>   -3.7462  -6.1247
#> 
#> (2,1,3,.,.) = 
#>  -0.9526  1.6430
#>   0.7569  2.2576
#>  -8.5560 -9.7035
#>   4.1302 -0.5691
#>  -6.7050 -6.3667
#> 
#> (3,1,3,.,.) = 
#>   0.2567  3.3491
#>  -8.7131  4.7589
#>   1.5954 -1.4664
#>  -5.7033  2.4085
#>   0.6070 -5.4705
#> 
#> (1,2,3,.,.) = 
#>  -5.4380 -2.4121
#>  -5.0166 -2.6143
#>  -4.2476  7.5749
#>  -1.9120  1.3121
#>   3.9339  0.7290
#> 
#> (2,2,3,.,.) = 
#>    4.8919  -9.6614
#>   -1.7365  -6.4366
#>    0.4799  10.6237
#>    3.4599   0.6882
#>   -3.5629  -1.7692
#> 
#> (3,2,3,.,.) = 
#>  -1.9953  2.9489
#>   0.9656 -4.4180
#>  -5.9864  0.6757
#>   1.9730 -1.4202
#>  -3.7969  5.7649
#> 
#> (1,3,3,.,.) = 
#>  -7.5540  0.7540
#>   1.8231 -0.1723
#>   2.5023  7.4179
#>  -0.1313  2.8180
#>   0.8127 -5.8250
#> 
#> (2,3,3,.,.) = 
#>  -8.2056 -0.8482
#>   2.0683 -18.6686
#>  -2.5826  2.1679
#>  -1.5549 -12.3176
#>   6.6992  0.0094
#> 
#> (3,3,3,.,.) = 
#>  -2.8456 -1.2566
#>  -14.1733 -0.3817
#>   2.1714 -3.5512
#>   2.5057 -5.7774
#>  -1.3546 -0.9014
#> 
#> (1,1,4,.,.) = 
#>  -3.1630 -3.0368
#>  -3.0379  2.4987
#>  -9.1058 -4.3333
#>   2.5158 -2.3408
#>   3.3329  7.4307
#> 
#> (2,1,4,.,.) = 
#>  -4.2199 -8.1722
#>   3.2436  4.1109
#>  -6.8430 -0.4683
#>   4.4940  2.2769
#>   4.1490  1.4424
#> 
#> (3,1,4,.,.) = 
#>  -16.3552   2.9946
#>    3.5135 -11.1236
#>   -8.3821   1.4669
#>    4.1577  -7.4396
#>   -2.0552  -8.3233
#> 
#> (1,2,4,.,.) = 
#>   2.3009  0.8086
#>   2.9567  7.0439
#>  -5.5452  7.5589
#>   4.3938 -2.3450
#>   0.4041 -2.7073
#> 
#> (2,2,4,.,.) = 
#>  -4.7550  3.1303
#>  -7.0661  5.5174
#>  -0.7417 -3.8804
#>   0.9515  2.4527
#>   2.2833  3.4872
#> 
#> (3,2,4,.,.) = 
#>  -0.3079  0.4233
#>  -6.6129  1.8303
#>   4.4564 -10.3904
#>  -5.6722  0.7938
#>  -10.3964  3.9085
#> 
#> (1,3,4,.,.) = 
#>  -5.9335  0.5387
#>   2.9040  1.9196
#>   1.6259 -1.0794
#>  -5.7042  0.2070
#>   2.3524  8.7244
#> 
#> (2,3,4,.,.) = 
#>   1.3367  0.6971
#>  -6.6217 -2.8865
#>   6.6995  8.9511
#>   7.3736  1.8491
#>  -3.0822  2.3404
#> 
#> (3,3,4,.,.) = 
#>    1.8480   6.0885
#>   -3.5436  -3.4685
#>   -2.1250  13.5269
#>   -3.5195   2.9393
#>    3.3513   2.1208
#> 
#> (1,1,5,.,.) = 
#>   -3.5185  -5.6899
#>    1.9912  -2.4328
#>   -5.4953  -5.4816
#>   -1.9472   1.7620
#>   10.5903   1.6089
#> 
#> (2,1,5,.,.) = 
#>   5.3416 -2.4417
#>  -3.3630 -0.8902
#>   4.4654  0.5536
#>  -6.2983  4.7187
#>  -2.3659 -6.3330
#> 
#> (3,1,5,.,.) = 
#>   4.5212 -2.4675
#>  -1.9087 -6.6206
#>  -0.2901 -14.0561
#>   2.3330 -0.0615
#>   7.6664  3.0468
#> 
#> (1,2,5,.,.) = 
#>  -0.9788  1.4510
#>   2.5766 -1.1039
#>  -6.9245 -3.7907
#>  -3.4216 -2.8843
#>   0.9847  5.5752
#> 
#> (2,2,5,.,.) = 
#>  -1.5066 -1.8273
#>  -2.2659 -6.0669
#>  -7.6243  3.6512
#>   0.9468  1.5735
#>  -2.6285 -2.4254
#> 
#> (3,2,5,.,.) = 
#>   0.9890  4.6629
#>  -3.8570 -2.8574
#>   2.9421 -5.6957
#>  -5.5224 -7.0839
#>   1.6589  5.8257
#> 
#> (1,3,5,.,.) = 
#>   -4.6225  -4.3016
#>    3.8103  15.1545
#>   11.3303   8.9338
#>   -6.1341  10.5460
#>    1.0033  -7.8533
#> 
#> (2,3,5,.,.) = 
#>  -5.7470 -3.2749
#>  -7.8943  2.6412
#>  -5.5028 -1.3226
#>   5.4086  3.1606
#>   2.9214 -4.6449
#> 
#> (3,3,5,.,.) = 
#>  -2.7212 -6.4866
#>  -2.5019 -7.4684
#>   0.3702 -2.9774
#>   7.1879 -0.6440
#>  -1.9355 -2.9521
#> [ CPUFloatType{3,3,5,5,2} ]