Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   5.1926 -0.1805
#>  -3.9580  5.8640
#>   6.2761 -3.2430
#>   9.8595  1.1676
#>   5.5408 -2.3278
#> 
#> (2,1,1,.,.) = 
#>   7.2263  1.2131
#>   3.5034  8.2771
#>   3.1381  7.6635
#>   3.4411  5.3335
#>  -2.3511 -0.1368
#> 
#> (3,1,1,.,.) = 
#>  -9.7348  6.6071
#>   3.2657 -2.5957
#>  -3.9301  5.3209
#>   0.2248 -5.5096
#>  -0.2995  3.3568
#> 
#> (1,2,1,.,.) = 
#>    0.1904   2.5589
#>    0.4552  -4.0384
#>    1.6734 -13.7100
#>    1.3367   6.2746
#>   10.1962   7.3541
#> 
#> (2,2,1,.,.) = 
#>   3.0278 -2.1266
#>  -5.2291  7.3109
#>   1.8790  6.5279
#>   2.7635 -2.4878
#>   5.1568 -12.3658
#> 
#> (3,2,1,.,.) = 
#>  -5.5423  2.2701
#>  -2.1105  6.0503
#>  -4.8682  3.9837
#>   0.5938 -2.3502
#>  -0.8288  3.4735
#> 
#> (1,3,1,.,.) = 
#>  -0.0774  2.1433
#>   1.1713 -3.5081
#>  -10.2624 -0.4668
#>   5.7724 -4.4691
#>  -3.5676  5.5122
#> 
#> (2,3,1,.,.) = 
#>   -2.9706   3.6636
#>   -4.8229  -1.2184
#>   -3.4403  -1.5138
#>   -0.4853  -2.3022
#>    4.0851  10.2229
#> 
#> (3,3,1,.,.) = 
#>  -1.2583  6.6607
#>  -7.5438  1.3889
#>   6.1397 -4.7388
#>  -3.7965  3.5392
#>  -5.9447 -0.4572
#> 
#> (1,1,2,.,.) = 
#>   7.1610  6.6054
#>  -1.1348  1.5555
#>   3.5484 -0.4778
#>  -1.4849  6.6081
#>  -0.5481  4.0515
#> 
#> (2,1,2,.,.) = 
#>  -4.0054  2.8158
#>   4.2154  5.1373
#>   1.2833  8.2760
#>   3.7217 -8.6486
#>  -1.4937 -5.2129
#> 
#> (3,1,2,.,.) = 
#>  -1.0042 -1.1464
#>   4.8894 -6.4975
#>  -0.8148  0.2599
#>  -4.9751  1.9764
#>  -6.0878 -0.9118
#> 
#> (1,2,2,.,.) = 
#>    1.5631   8.4873
#>    1.7452   1.2358
#>   -4.4967   0.6604
#>   -0.1634  10.6987
#>    4.8851  -0.9044
#> 
#> (2,2,2,.,.) = 
#>   6.7873  1.9846
#>   5.9852 -2.1296
#>   3.0334 -7.6664
#>   1.2412  6.0806
#>   8.6393 -2.4906
#> 
#> (3,2,2,.,.) = 
#>   3.3793  5.4424
#>   1.2192  4.9083
#>   3.1422 -1.4830
#>  -12.8136  0.9918
#>  -4.5165  1.1942
#> 
#> (1,3,2,.,.) = 
#>  -0.7836 -13.7282
#>  -7.7224 -3.3258
#>  -1.8579 -3.9250
#>   6.5842 -7.2296
#>   6.1606  1.3222
#> 
#> (2,3,2,.,.) = 
#>   0.1385 -8.2769
#>  -12.5008 -10.8349
#>  -5.1415  5.6843
#>   2.2185 -7.5096
#>  -6.2111 -1.4911
#> 
#> (3,3,2,.,.) = 
#>   3.8858 -9.3698
#>  -2.3747  4.8237
#>  -0.1514 -8.7229
#>  -0.4831  4.9851
#>   0.8111 -5.8530
#> 
#> (1,1,3,.,.) = 
#>   3.0339  4.8418
#>   4.5921 -2.2415
#>   2.3554 -3.3354
#>  -4.3658  2.1412
#>  -1.5655 -4.6739
#> 
#> (2,1,3,.,.) = 
#>  -1.9613 -1.5487
#>   2.4914 -5.9635
#>  -2.4649  4.9279
#>  -1.3478  1.2485
#>   8.3236 -4.7375
#> 
#> (3,1,3,.,.) = 
#>   12.2607  -3.4063
#>   -3.6919   5.6668
#>    6.1291  -4.9187
#>   -0.9711   0.3730
#>    0.8944  -2.3152
#> 
#> (1,2,3,.,.) = 
#>  -13.5588   5.3589
#>    4.3759  -1.5228
#>  -10.0839   0.3581
#>    0.5823   3.5901
#>   -7.9368  -3.9118
#> 
#> (2,2,3,.,.) = 
#>   1.1660  8.1805
#>   1.0622 -3.2256
#>  -0.0829  1.0276
#>   5.4565 -9.4576
#>   8.0999 -6.1543
#> 
#> (3,2,3,.,.) = 
#>  -0.7362 -4.3615
#>   0.3562 -1.0177
#>  -7.5982 -8.7456
#>   2.5468  3.4394
#>   2.3971  0.5053
#> 
#> (1,3,3,.,.) = 
#>   -2.0028  -3.7111
#>    6.1333   4.4754
#>   -5.5886 -10.3001
#>    9.3364  10.2133
#>   -4.0977  -1.0885
#> 
#> (2,3,3,.,.) = 
#>   1.1104 -7.1246
#>  -0.7187 -2.6510
#>   9.7203 -9.4094
#>   5.0322 -2.9585
#>  -2.4715  5.6305
#> 
#> (3,3,3,.,.) = 
#>  -0.3292 -4.8204
#>   2.2378 -0.6516
#>   7.1541 -2.1031
#>  -0.0056 -6.6863
#>   5.0521 -0.1263
#> 
#> (1,1,4,.,.) = 
#>   8.7732 -4.0209
#>  -4.5046 -2.6968
#>  -2.8980 -5.3990
#>   6.7367  5.0867
#>   2.7935  0.6667
#> 
#> (2,1,4,.,.) = 
#>   1.1646 -5.6612
#>  -6.3464  8.2441
#>   6.7291  3.7418
#>   0.7084  5.4624
#>   1.9256 -3.0797
#> 
#> (3,1,4,.,.) = 
#>  -2.4986 -1.9561
#>   4.1193  2.0918
#>  -7.1081 -1.4686
#>   0.9545 -10.4444
#>  -2.8774  6.6328
#> 
#> (1,2,4,.,.) = 
#>   2.1641  1.3522
#>   0.9901 -2.1287
#>   4.9534 -0.1487
#>   6.7989 -1.8758
#>  -3.2206 -14.3547
#> 
#> (2,2,4,.,.) = 
#>  -4.6182 -2.9769
#>   4.9982 -4.7885
#>   1.8287  2.2206
#>  -11.2709  1.7415
#>   1.4153  2.9516
#> 
#> (3,2,4,.,.) = 
#>  -9.5389  5.2294
#>  -6.4391 -0.7533
#>  -4.8627 -3.6128
#>  -7.6494 -5.3012
#>  -5.4864 -1.4717
#> 
#> (1,3,4,.,.) = 
#>    2.2533  11.7702
#>   -0.9739  -1.4599
#>    7.2885   5.1207
#>   -0.6666  -3.4859
#>   -2.4133  -9.1130
#> 
#> (2,3,4,.,.) = 
#>  -2.4365 -3.7746
#>   6.8885 -2.6690
#>   4.7459  7.5945
#>   5.6403 -5.7786
#>   6.2713  1.0531
#> 
#> (3,3,4,.,.) = 
#>  -1.0296  4.7289
#>   8.6256  0.8473
#>  -2.2628  5.6987
#>   1.6552  2.3149
#>  -2.6575  1.3828
#> 
#> (1,1,5,.,.) = 
#>    4.1782  -0.8189
#>    8.7975  -6.4781
#>   11.8341   0.3960
#>   -1.6627  -1.7055
#>   -3.3477  -4.8645
#> 
#> (2,1,5,.,.) = 
#>  -10.5788   9.9898
#>   -0.9949  -0.9921
#>    3.2661  -7.1057
#>   10.7822   1.4921
#>   -7.0833  -5.0104
#> 
#> (3,1,5,.,.) = 
#>   5.1259  3.0363
#>   1.0258 -1.2629
#>  -4.7300 -4.9555
#>  -1.8618  0.6167
#>   0.0608 -2.5222
#> 
#> (1,2,5,.,.) = 
#>  -5.7764  0.3446
#>  -2.6591  0.6130
#>   1.8209 -2.4371
#>   3.9651  3.4221
#>  -8.2373  2.6622
#> 
#> (2,2,5,.,.) = 
#>   1.2864 -0.6195
#>  -3.4328  8.8002
#>   1.2390 -6.0402
#>   3.2483  0.3766
#>  -1.7587  8.7192
#> 
#> (3,2,5,.,.) = 
#>   0.7558  0.9970
#>   0.3780 -10.1833
#>  -4.9043  4.4491
#>   1.2091  1.1090
#>   3.1670 -3.1875
#> 
#> (1,3,5,.,.) = 
#>   1.2229  5.1120
#>  -2.4024  1.7183
#>   1.4078 -11.0723
#>  -4.5581 -2.5037
#>   5.6273  5.3186
#> 
#> (2,3,5,.,.) = 
#>   8.4301  5.6034
#>   3.7579 -0.1479
#>  -0.8292  0.0071
#>  -2.1075 -1.6596
#>   6.1541 -2.6591
#> 
#> (3,3,5,.,.) = 
#>  -1.8733  6.6772
#>   1.8630 -2.6401
#>  -2.4881 -10.1504
#>  -6.4280 -4.6207
#>   3.2086 -7.4940
#> [ CPUFloatType{3,3,5,5,2} ]