Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>    3.6671  -2.2888
#>    4.4099  -5.1085
#>   -7.7229   9.5636
#>   -5.4345  14.8604
#>   -1.3873   3.8088
#> 
#> (2,1,1,.,.) = 
#>   1.0926 -2.7212
#>   0.4558  6.1655
#>   5.0687 -7.5283
#>  -0.0500 -4.7484
#>   1.6789 -1.0975
#> 
#> (3,1,1,.,.) = 
#>   5.2651  5.8184
#>  -5.1943 -2.2855
#>   5.0994 -2.3146
#>   4.9735  3.7482
#>  -4.1747  2.2076
#> 
#> (1,2,1,.,.) = 
#>   7.5999  1.9453
#>  -2.1289  3.0765
#>  -0.4013  3.7361
#>   3.8384 -10.8576
#>  -2.3666 -2.8882
#> 
#> (2,2,1,.,.) = 
#>  -2.9245 -5.6171
#>   5.4342 -0.0945
#>  -2.8909  2.3308
#>   8.9067  7.2946
#>   3.7510 -5.7639
#> 
#> (3,2,1,.,.) = 
#>  -8.3544  0.8636
#>  -3.9031 -0.6621
#>   2.5023  2.3250
#>  -2.6630 -5.9077
#>   0.4692  1.9862
#> 
#> (1,3,1,.,.) = 
#>   5.9041  5.9765
#>  -12.6200  0.8591
#>  -1.4055 -5.8535
#>   3.2980  1.5478
#>   1.5017 -0.9507
#> 
#> (2,3,1,.,.) = 
#>   8.8915 -7.0543
#>  -3.7091 -5.5742
#>   5.8849 -0.8833
#>  -5.4497 -1.8255
#>  -0.4735 -5.0624
#> 
#> (3,3,1,.,.) = 
#>  -0.3435 -2.8270
#>   0.9013 -10.6026
#>   4.4843 -4.3989
#>  -9.5389  6.3674
#>   3.0468 -7.7719
#> 
#> (1,1,2,.,.) = 
#>  -5.6204 -0.7920
#>  -4.4550 -2.7015
#>   5.5209 -3.3396
#>  -4.1520 -2.8881
#>   2.1152 -9.4264
#> 
#> (2,1,2,.,.) = 
#>   4.5204 -5.4533
#>   1.3830 -10.0942
#>  -3.5677  6.5747
#>   7.6110 -5.6209
#>  -0.7978 -5.4640
#> 
#> (3,1,2,.,.) = 
#>  -1.0438 -10.0683
#>  -7.3108  9.6545
#>   5.7408  4.5261
#>   5.7892 -0.2534
#>   0.6549  1.3158
#> 
#> (1,2,2,.,.) = 
#>  -10.9971   5.7069
#>   -2.7816   1.3359
#>    2.2380   6.7000
#>   16.3338   9.1275
#>    6.2775   4.7238
#> 
#> (2,2,2,.,.) = 
#>  -1.0188 -0.1418
#>  -3.0921  9.7907
#>  -2.3654  1.7503
#>  -0.6421 -3.1597
#>  -2.0618  1.6024
#> 
#> (3,2,2,.,.) = 
#>  -1.6726  9.6023
#>  -5.5306 -3.7222
#>  -5.2006 -0.2981
#>   3.0360  4.5706
#>   4.5857  3.4093
#> 
#> (1,3,2,.,.) = 
#>   2.3589 -6.9115
#>   9.4954 -5.8078
#>  -13.0843  1.8405
#>  -0.3458  6.1455
#>  -8.0800 -2.2081
#> 
#> (2,3,2,.,.) = 
#>   5.1509  6.4474
#>   4.4580  8.8249
#>   0.6111 -5.2935
#>  -6.7116 -6.3294
#>  -4.5872  4.1418
#> 
#> (3,3,2,.,.) = 
#>    3.6895  -0.2323
#>   -4.6699  -5.4480
#>    0.6264   0.5072
#>    4.2964  14.5822
#>   -0.2023  -9.6320
#> 
#> (1,1,3,.,.) = 
#>   -0.9455   6.9554
#>   11.2597  -5.2245
#>    1.0645   2.9332
#>   11.6882  -9.8896
#>    4.8159   3.2283
#> 
#> (2,1,3,.,.) = 
#>   2.2072  1.8503
#>   2.1708  5.3847
#>   7.9692 -0.2485
#>   7.1178 -3.0291
#>  -2.0943 -6.8455
#> 
#> (3,1,3,.,.) = 
#>   -4.4625  -6.0829
#>    4.1856  -0.5919
#>    3.5106   5.0807
#>   -2.6306  -2.3954
#>   -3.7941  18.4265
#> 
#> (1,2,3,.,.) = 
#>   2.9953  0.7424
#>   5.8017  8.0116
#>  -9.2276 -0.3576
#>   8.8628 -1.2618
#>   0.9937 -5.3008
#> 
#> (2,2,3,.,.) = 
#>   1.8452 -3.4052
#>  -6.2886 -6.2817
#>  -2.1137 -2.1591
#>  -2.7066  2.2777
#>  -7.3123 -3.9195
#> 
#> (3,2,3,.,.) = 
#>   15.0349   5.3844
#>    5.9374 -11.8243
#>   10.4297  -8.6687
#>   -8.7371   0.4748
#>   -0.9313   5.1286
#> 
#> (1,3,3,.,.) = 
#>   1.7697 -1.2500
#>  -3.5727  0.3612
#>  -2.0708 -5.6266
#>  -0.6476 -3.7022
#>  -3.4383  1.0280
#> 
#> (2,3,3,.,.) = 
#>   3.9025 -3.3036
#>   7.8919 -6.3574
#>  -2.2269 -7.5253
#>  -9.2310  0.8579
#>   2.1639  6.2846
#> 
#> (3,3,3,.,.) = 
#>   -0.9424   5.7620
#>   -3.6371  -2.5944
#>   -4.5043   3.0559
#>   13.8303   1.2080
#>  -10.2492   3.0194
#> 
#> (1,1,4,.,.) = 
#>   0.4850  7.0115
#>   0.7275  2.9707
#>   0.5919 -11.6151
#>   9.0668  4.0814
#>   1.1337 -5.0898
#> 
#> (2,1,4,.,.) = 
#>   3.3460 -0.9591
#>  -3.2343 -4.7261
#>  -2.0054 -1.4381
#>   6.3444 -3.6104
#>  -1.8211  0.0306
#> 
#> (3,1,4,.,.) = 
#>   -0.8868   1.7885
#>   14.1786   6.6297
#>   -5.4119   0.4432
#>   -7.7246  -3.2153
#>   -7.5184  -2.2390
#> 
#> (1,2,4,.,.) = 
#>  -3.3477  6.9556
#>   3.2198 -3.8248
#>  -1.1144 -0.2870
#>  -7.4756  5.0291
#>  -10.8763  2.2239
#> 
#> (2,2,4,.,.) = 
#>   1.2369 -0.9711
#>   1.8564 -12.0556
#>   4.6312  8.5101
#>  -2.3434 -3.0990
#>  -3.7365 -0.6163
#> 
#> (3,2,4,.,.) = 
#>  -5.4437 -1.7526
#>  -3.9453  1.9468
#>   2.0589  4.5907
#>   3.5087 -5.1658
#>  -1.3342 -1.3016
#> 
#> (1,3,4,.,.) = 
#>  -6.9401  8.0024
#>   5.9500  2.4279
#>  -4.3024 -3.1429
#>   8.5838 -2.4898
#>  -7.4856 -4.8371
#> 
#> (2,3,4,.,.) = 
#>  -8.7625 -1.0364
#>  -5.5840 -2.2864
#>  -2.8800  0.6786
#>  -3.9934 -0.2251
#>   7.2014  0.3872
#> 
#> (3,3,4,.,.) = 
#>   9.0802 -0.5793
#>   1.0460  8.0878
#>   3.0548 -4.2302
#>  -4.0991 -1.3646
#>  -4.8712 -8.7084
#> 
#> (1,1,5,.,.) = 
#>   -6.3219  -4.0132
#>    3.6285   5.5858
#>    0.1910  -7.5489
#>   11.6085   5.7053
#>   -2.1208   0.8132
#> 
#> (2,1,5,.,.) = 
#>   4.7186 -4.1521
#>  -2.4429 -1.3773
#>  -8.4875  0.6946
#>  -3.2214  0.7267
#>   1.1814  2.0536
#> 
#> (3,1,5,.,.) = 
#>    9.5409  -3.6399
#>    2.5442  -3.6181
#>   15.1822  -1.7984
#>   -1.3709  -9.6450
#>    6.7756  -3.7542
#> 
#> (1,2,5,.,.) = 
#>  -4.7525 -5.0711
#>   0.1986 -3.9157
#>  -13.0298  4.7490
#>  -7.8589  0.8955
#>   6.3688  3.6269
#> 
#> (2,2,5,.,.) = 
#>  -7.1350  9.7280
#>  -2.6205  5.9598
#>  -3.0520 -4.1621
#>   5.2045 -3.6286
#>  -15.3845 -5.2054
#> 
#> (3,2,5,.,.) = 
#>  -4.2608  9.9626
#>   4.9595  0.0484
#>  -0.8592 -6.8057
#>   4.6898 -4.2517
#>   4.9115 -4.8788
#> 
#> (1,3,5,.,.) = 
#>  -2.7294  6.0444
#>   8.0547 -3.3578
#>   1.4715  1.5683
#>  -1.7881  3.1560
#>   1.6718 -2.0340
#> 
#> (2,3,5,.,.) = 
#>  -3.8268 -8.5304
#>   6.1499  6.3992
#>  -2.2179  2.6013
#>   1.3980  4.4650
#>   4.3426 -6.2383
#> 
#> (3,3,5,.,.) = 
#>    2.6879   1.7470
#>    6.7903   4.7292
#>   -0.6413   5.4014
#>   12.7039   1.2500
#>   -5.9588  -2.2582
#> [ CPUFloatType{3,3,5,5,2} ]