Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -3.4949 -4.8215
#>  -2.4973  5.0641
#>   4.6545 -3.8295
#>  -11.3533  0.1460
#>  -3.9779  4.6840
#> 
#> (2,1,1,.,.) = 
#>   -5.5149  10.3641
#>    4.9081  -6.6062
#>   -2.6797  -2.0167
#>    7.9178  -2.3001
#>   -7.4346  -8.1156
#> 
#> (3,1,1,.,.) = 
#>   2.1732 -3.3206
#>   0.2538 -3.0920
#>  -7.4019  1.5851
#>   4.2973 -8.2244
#>  -3.9462  1.2542
#> 
#> (1,2,1,.,.) = 
#>  -10.7970  -2.1936
#>   -5.5555   0.7801
#>    1.4463  -7.7474
#>   10.4016  -1.8303
#>    5.0015   0.1997
#> 
#> (2,2,1,.,.) = 
#>   5.2114 -6.3596
#>  -5.4012  4.5391
#>   8.2644  0.4777
#>  -1.3582 -8.8912
#>   6.5195  6.3266
#> 
#> (3,2,1,.,.) = 
#>  -4.7936 -3.2643
#>  -8.2106 -1.3274
#>   3.2675 -3.9146
#>  -5.9933  0.2253
#>  -3.2496 -3.5943
#> 
#> (1,3,1,.,.) = 
#>    3.7965  -1.7185
#>   12.6888  -2.5721
#>   -3.0281  -3.3852
#>    0.4130   5.0324
#>    1.4185  -5.9977
#> 
#> (2,3,1,.,.) = 
#>  -5.7586 -0.0080
#>   3.3266 -1.4097
#>   2.6192  1.0723
#>   5.6441 -4.0345
#>  -1.0945 -2.5985
#> 
#> (3,3,1,.,.) = 
#>   9.2803 -6.7915
#>   2.7621  4.1956
#>  -9.1796 -0.2418
#>  -0.6762  3.0468
#>  -2.8360  3.0661
#> 
#> (1,1,2,.,.) = 
#>   9.4812  2.5372
#>  -8.7255 -0.0919
#>  -5.1414 -18.1980
#>  -3.8427  8.4672
#>   3.0343 -6.4412
#> 
#> (2,1,2,.,.) = 
#>  -7.7506  1.9099
#>  -3.9246  5.2812
#>   5.6748 -5.5835
#>  -6.8941 -7.6430
#>  -11.7988  1.9197
#> 
#> (3,1,2,.,.) = 
#>  -7.7257  7.5459
#>  -5.3805  6.5935
#>   6.5988  3.2902
#>  -8.7703  7.1086
#>   5.0612  4.3546
#> 
#> (1,2,2,.,.) = 
#>   -4.0013  -4.7121
#>   -2.7389  11.6706
#>   -0.8010   2.6811
#>    0.8314  -2.5209
#>   -3.2790   5.1226
#> 
#> (2,2,2,.,.) = 
#>   -6.1823  -0.5518
#>   -4.2654  -6.0731
#>    1.8598  -2.2522
#>   -1.5985  11.8313
#>    2.1323   0.4020
#> 
#> (3,2,2,.,.) = 
#>   1.6169 -0.1116
#>   7.2912  1.3512
#>  -4.7557 -6.0756
#>   8.5453 -3.0883
#>   5.7898 -3.7203
#> 
#> (1,3,2,.,.) = 
#>    3.1918   1.0034
#>    3.0115   0.7027
#>    7.2441  -3.8087
#>   -1.8648   3.9168
#>   15.0540  -1.5426
#> 
#> (2,3,2,.,.) = 
#>   0.8611 -3.7284
#>   2.1447  2.5566
#>   1.6607 -1.6911
#>  -6.3645 -4.0498
#>  -0.9116  3.0292
#> 
#> (3,3,2,.,.) = 
#>   0.7338  2.1984
#>  -6.7400 -4.5150
#>  -2.4759  4.8934
#>  -3.8730  3.4081
#>  -8.2080 -3.7466
#> 
#> (1,1,3,.,.) = 
#>  -0.1293  1.1297
#>   5.3101  5.1319
#>  -4.1501 -0.2873
#>   0.5792  3.5303
#>   4.1124  8.6162
#> 
#> (2,1,3,.,.) = 
#>   12.9861   7.7185
#>  -10.2129 -10.8062
#>    3.1812  -2.8618
#>    0.5844   4.7613
#>    7.8434   0.6273
#> 
#> (3,1,3,.,.) = 
#>   -1.4839   2.2398
#>   -1.7481  12.8087
#>    3.2694  -2.8049
#>    1.0853  -6.6928
#>   -4.4426   2.8810
#> 
#> (1,2,3,.,.) = 
#>  -3.1517 -2.0651
#>  -4.7524  1.4257
#>   0.4116  6.6381
#>   0.9421 -2.4183
#>   1.4123  5.4408
#> 
#> (2,2,3,.,.) = 
#>   1.3846 -0.6427
#>   0.2047  0.4547
#>  -3.5440  5.9846
#>   8.9318 -4.7878
#>   0.5765  1.1310
#> 
#> (3,2,3,.,.) = 
#>  -3.6107 -3.5215
#>   0.4796  1.0828
#>  -2.3944  5.0589
#>   4.5956  6.9434
#>   6.7883  0.3939
#> 
#> (1,3,3,.,.) = 
#>    1.3749  11.4946
#>   -0.9196  -1.3037
#>   -2.1022   2.5043
#>   -1.9710   6.1347
#>   -0.9441   6.1189
#> 
#> (2,3,3,.,.) = 
#>   1.6624  1.2452
#>   5.7359  2.8267
#>   6.1434  5.8009
#>  -2.8896 -2.8039
#>  -1.7474  6.2643
#> 
#> (3,3,3,.,.) = 
#>   -4.2999   3.7378
#>    4.1583   5.4148
#>  -13.8154   0.5522
#>    0.7290  -6.5571
#>   -0.1854  11.1064
#> 
#> (1,1,4,.,.) = 
#>  -0.4568 -7.5987
#>   3.5135  2.6966
#>  -4.5489 -0.9878
#>  -3.8029 -8.3286
#>  -2.4313 -5.1100
#> 
#> (2,1,4,.,.) = 
#>   1.9122 -0.2183
#>  -2.4762  3.0914
#>  -3.9079  3.3401
#>   1.8117 -6.8832
#>  -3.5119 -3.0414
#> 
#> (3,1,4,.,.) = 
#>   3.4869  4.4926
#>   0.6540 -1.9407
#>   2.0894 -2.1367
#>  -0.7582 -6.9929
#>   5.6574 -4.2861
#> 
#> (1,2,4,.,.) = 
#>   -6.0402  -2.4812
#>   -3.1719   6.4945
#>  -13.5427  -4.4467
#>   10.2270  -1.9620
#>    6.6959   5.5550
#> 
#> (2,2,4,.,.) = 
#>  -7.7121 -6.7875
#>   4.6616 -3.3959
#>  -1.1565  1.4736
#>   1.7819  7.5011
#>  -14.0101 -0.8870
#> 
#> (3,2,4,.,.) = 
#>    7.7750   1.3838
#>   -1.6488  -1.4005
#>    0.9341   2.2127
#>   -3.2065  10.0125
#>    5.1103 -10.9137
#> 
#> (1,3,4,.,.) = 
#>  -2.4762  3.1547
#>   7.9210  0.5620
#>  -1.9325 -1.0421
#>  -0.6271 -6.6022
#>   0.5003 -9.7472
#> 
#> (2,3,4,.,.) = 
#>   7.6838  1.6045
#>  -1.0383  8.8142
#>   0.0174  4.8284
#>  -2.9461  4.2845
#>  -4.8300 -2.8266
#> 
#> (3,3,4,.,.) = 
#>   1.6625 -10.2569
#>   6.7304 -6.6731
#>  -6.6115 -3.6964
#>   0.8436 -4.0048
#>  -7.4337 -5.3315
#> 
#> (1,1,5,.,.) = 
#>   6.4282 -1.6375
#>   3.7683  1.5584
#>   1.6466 -7.6858
#>  -6.8407 -4.6635
#>   7.7118  6.4480
#> 
#> (2,1,5,.,.) = 
#>  -2.7173  4.4824
#>  -8.8958 -3.2716
#>  -4.6480  2.2375
#>  -9.4683  8.4622
#>  -1.5731  9.1293
#> 
#> (3,1,5,.,.) = 
#>   11.0640  -1.2049
#>    1.0735   4.0419
#>   -5.1843   0.7924
#>   -6.8139   4.3204
#>   -4.5244   5.4722
#> 
#> (1,2,5,.,.) = 
#>   -3.0205   4.8499
#>    2.7368   5.8968
#>    1.6684  -1.3006
#>   12.9055  -6.8439
#>    1.1834  -2.9593
#> 
#> (2,2,5,.,.) = 
#>   11.0264  -4.2025
#>    0.2907  -9.1654
#>    4.0701   2.9439
#>   -2.7229   2.9145
#>   -4.2815  -1.3990
#> 
#> (3,2,5,.,.) = 
#>   -6.1083  -0.0102
#>    0.8060   1.4569
#>    0.3655   1.4686
#>    1.2789   8.3308
#>   -3.2136  10.1796
#> 
#> (1,3,5,.,.) = 
#>   7.4817  0.0735
#>   2.0761 -6.8251
#>  -9.2858 -3.4763
#>  -6.6524 -8.5773
#>  -10.2074  2.5098
#> 
#> (2,3,5,.,.) = 
#>    0.1287   1.4598
#>   -3.7679   2.4741
#>   -2.0110   7.1209
#>   -9.7437  -1.3532
#>    5.8457  11.2920
#> 
#> (3,3,5,.,.) = 
#>   -1.2036  -1.7655
#>   -9.5918  -3.4374
#>   -7.8209  10.1298
#>    1.3535   4.8356
#>    3.9271   3.2610
#> [ CPUFloatType{3,3,5,5,2} ]