Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -5.8855  7.4005
#>   1.7775 -6.4448
#>  -1.7210 -10.3100
#>  -5.8832  2.7778
#>  -1.8390 -1.0214
#> 
#> (2,1,1,.,.) = 
#>   4.3284 -8.1251
#>   2.1803 -3.4741
#>   6.4840  1.4323
#>   0.7880  0.8525
#>   3.0858  3.8159
#> 
#> (3,1,1,.,.) = 
#>   -9.1000  12.3169
#>   -4.3602  -4.2630
#>   -2.2735   3.3649
#>   -0.5892 -10.1999
#>   -8.7068   0.4698
#> 
#> (1,2,1,.,.) = 
#>   2.6392 -3.7288
#>   0.4197  4.6963
#>  -6.5447  3.8678
#>  -2.5554  5.4050
#>   0.0816 -2.2727
#> 
#> (2,2,1,.,.) = 
#>  -2.0464 -0.9336
#>  -1.6276  0.5518
#>  -0.9872  1.1789
#>  -5.6243  9.4194
#>   0.8287 -0.3502
#> 
#> (3,2,1,.,.) = 
#>   4.1943  0.1552
#>   0.3348  0.7264
#>  -3.8656 -1.1679
#>   8.7895 -2.7773
#>   5.3489 -4.9609
#> 
#> (1,3,1,.,.) = 
#>   1.9893  1.0913
#>  -1.7545 -3.3203
#>   0.2116  2.9216
#>  -4.1916  4.2554
#>   3.3651 -3.3526
#> 
#> (2,3,1,.,.) = 
#>   3.6504 -3.6774
#>  -5.4097 -6.8980
#>   1.3195  2.4102
#>   2.7191  0.9983
#>   3.3955 -0.1388
#> 
#> (3,3,1,.,.) = 
#>  -0.5545  1.7328
#>  -4.7732 -9.3936
#>  -1.9649 -0.6531
#>  -6.9821  6.1099
#>  -5.6714 -0.7837
#> 
#> (1,1,2,.,.) = 
#>  -4.7680 -4.7571
#>   7.8038 -1.4204
#>   2.0586 -0.5978
#>  -5.3043  3.2700
#>   2.5967  2.2490
#> 
#> (2,1,2,.,.) = 
#>  -0.5863 -1.5090
#>  -3.2688 -7.4816
#>   3.1590  4.8414
#>  -1.8445  4.7095
#>   2.4126  3.1769
#> 
#> (3,1,2,.,.) = 
#>   0.8267 -2.3123
#>  -4.5416  5.1965
#>  -6.2749  3.8102
#>   2.1419  0.5003
#>  -1.4244  9.5632
#> 
#> (1,2,2,.,.) = 
#>  -4.2917 -2.8764
#>   0.8687  1.1856
#>   3.4443  0.2072
#>  -2.5774  3.1973
#>  -2.3415 -0.2988
#> 
#> (2,2,2,.,.) = 
#>   5.7809  1.2024
#>  -2.9511 -2.7794
#>   2.5481  0.1741
#>  -1.9356 -2.2712
#>  -5.4295 -5.7757
#> 
#> (3,2,2,.,.) = 
#>   -4.3356  -0.9072
#>    4.4486  -7.3099
#>   -2.1787  10.2380
#>   -2.5954 -14.6892
#>   -8.6026  -3.0728
#> 
#> (1,3,2,.,.) = 
#>  -3.7558 -0.4780
#>   3.8941  4.3441
#>  -1.6606 -0.6364
#>   4.6687 -1.3336
#>   1.5200  5.3838
#> 
#> (2,3,2,.,.) = 
#>  -0.7894  1.5720
#>   0.0777 -2.9127
#>   0.6089 -3.7833
#>   3.1948 -3.9109
#>   1.3483 -1.4062
#> 
#> (3,3,2,.,.) = 
#>   7.1462 -10.1185
#>   1.0419  3.5681
#>   1.2407 -6.0569
#>  -0.0621  2.7892
#>   6.4209 -2.5629
#> 
#> (1,1,3,.,.) = 
#>  -7.2798 -0.9959
#>  -4.9036  4.9087
#>   6.1714 -3.0309
#>  -3.7971 -5.8232
#>   2.8407 -2.3883
#> 
#> (2,1,3,.,.) = 
#>  -4.9259 -0.1600
#>   0.7360 -9.7565
#>   2.8953 -2.0291
#>   0.4825 -4.8593
#>   0.7895  4.3277
#> 
#> (3,1,3,.,.) = 
#>  -2.4139  2.1933
#>  -0.2229  4.7885
#>   3.3573  7.9348
#>   2.3537  0.3774
#>  -5.6405 -5.1602
#> 
#> (1,2,3,.,.) = 
#>   3.3485 -6.0011
#>  -8.1558 -9.2352
#>  -3.1920 -6.2615
#>  -3.9924  5.0729
#>  -1.8094 -1.8742
#> 
#> (2,2,3,.,.) = 
#>    6.7169  -1.0789
#>    4.7814  10.9677
#>    3.2449   3.1941
#>  -10.1638  -9.6475
#>   -8.0627  -2.6224
#> 
#> (3,2,3,.,.) = 
#>    1.0571  -4.2510
#>   10.5263   0.2544
#>    9.7935   0.7411
#>   -4.3672  -1.6470
#>   -5.4007   2.7234
#> 
#> (1,3,3,.,.) = 
#>  -2.9205 -3.9642
#>  -0.3255  1.3595
#>   5.5339 -4.5481
#>   0.3170  9.0304
#>   1.3702  2.6969
#> 
#> (2,3,3,.,.) = 
#>   3.6284  0.7402
#>   4.1637 -1.2220
#>  -6.8070  7.2331
#>  -8.2203  1.1379
#>   5.9901 -5.6730
#> 
#> (3,3,3,.,.) = 
#>   0.5426  0.9300
#>  -4.3119  6.3104
#>  -3.4984 -8.6285
#>  -2.5132 -1.8755
#>  -2.0898  0.8301
#> 
#> (1,1,4,.,.) = 
#>   5.5397  6.9253
#>  -9.0711 -6.6941
#>  -3.9961  4.6210
#>  -2.0592  2.4409
#>  -2.4592 -0.0081
#> 
#> (2,1,4,.,.) = 
#>  -12.3905  -0.2830
#>   -3.0186   4.9168
#>    2.8164  -2.5051
#>    4.0393  -6.6592
#>   -4.4206  -3.6855
#> 
#> (3,1,4,.,.) = 
#>  -6.2867 -4.1337
#>  -0.8913 -0.9244
#>   1.7035  0.8646
#>  -1.8074  2.1040
#>   2.2625  1.3519
#> 
#> (1,2,4,.,.) = 
#>    0.3237  -0.9336
#>   12.1522   0.7372
#>    1.9278  -2.4831
#>   -0.0562   2.5630
#>   -1.0809   3.5710
#> 
#> (2,2,4,.,.) = 
#>   0.1559 -3.9282
#>  -6.2424  2.2199
#>  -1.9298  0.4852
#>   1.7141  9.5448
#>  -0.1504  3.0320
#> 
#> (3,2,4,.,.) = 
#>   -6.9496  -1.5572
#>   13.7681  11.2592
#>    0.8908  -1.1722
#>   -1.9739  -8.0395
#>   -4.8062   1.6422
#> 
#> (1,3,4,.,.) = 
#>   -6.1458  10.0672
#>  -11.2188   2.2893
#>    2.2591   2.5540
#>   -0.9200  -0.8934
#>   14.2071   3.0450
#> 
#> (2,3,4,.,.) = 
#>   5.7211 -3.0790
#>  -4.3222  4.6937
#>  -1.9580 -11.4381
#>   1.5498  0.1623
#>   2.8182  6.2636
#> 
#> (3,3,4,.,.) = 
#>    1.1626  -6.1698
#>   -9.9772   6.5304
#>   -4.1039  12.0264
#>   -5.4156  -6.4355
#>   -6.4437  -0.0080
#> 
#> (1,1,5,.,.) = 
#>  -11.8227  -3.2114
#>  -12.2429   3.1132
#>   -5.0612   2.3612
#>    1.1468   5.1537
#>   -0.5849   1.6175
#> 
#> (2,1,5,.,.) = 
#>  -7.6166  0.4313
#>   2.0231 -6.1409
#>  -0.2437 -3.5538
#>   5.4344 -2.5658
#>   1.4025 -0.8720
#> 
#> (3,1,5,.,.) = 
#>   1.2010 -7.9388
#>   6.9298  5.2968
#>   4.5879 -4.9498
#>  -0.8697  7.1597
#>  -11.9514  4.1636
#> 
#> (1,2,5,.,.) = 
#>  -4.5835 -0.4383
#>   0.3735  0.4568
#>  -13.7316  0.6499
#>   1.7335 -2.3985
#>   5.0373 -2.5320
#> 
#> (2,2,5,.,.) = 
#>  -5.6017  7.2184
#>  -1.9493 -1.5905
#>   0.6656  0.7146
#>  -5.3147 -8.5443
#>   2.9132 -12.8823
#> 
#> (3,2,5,.,.) = 
#>    3.5247  -4.5173
#>   -0.8144   4.9407
#>   13.2646   2.4973
#>    5.4050   7.8953
#>    0.8448   0.2405
#> 
#> (1,3,5,.,.) = 
#>  -4.5890 -2.3624
#>  -0.9465 -2.0590
#>  -4.9582 -0.3306
#>   1.0891  0.4830
#>   1.4687 -2.4643
#> 
#> (2,3,5,.,.) = 
#>   2.9715 -4.8153
#>   2.1618  0.9600
#>   1.7951 -4.5031
#>  -0.6807  3.1319
#>  -1.1816 -3.3113
#> 
#> (3,3,5,.,.) = 
#>   3.3843 -1.8101
#>  -9.1549 -7.0195
#>   4.9218 -4.8158
#>  -4.5664 -7.4126
#>   0.5907  3.4384
#> [ CPUFloatType{3,3,5,5,2} ]