Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   3.5648  3.3506
#>  -0.2250 -1.8878
#>  -1.7832  3.7417
#>  -1.0781  3.1161
#>   9.0569 -3.2340
#> 
#> (2,1,1,.,.) = 
#>   7.8628  5.0856
#>  -6.0213 -3.8247
#>  -5.1438  4.2758
#>  -3.8698 -8.2088
#>   6.6028 -4.0209
#> 
#> (3,1,1,.,.) = 
#>   5.1998  3.8359
#>  -11.3599  3.9293
#>   0.5762 -0.8991
#>   4.5552 -0.2311
#>   0.9835 -3.7703
#> 
#> (1,2,1,.,.) = 
#>  -9.8061 -2.9487
#>   6.8663  6.5270
#>  -2.7469  2.4229
#>  -0.8377 -4.0692
#>  -1.9348  0.3804
#> 
#> (2,2,1,.,.) = 
#>  -0.7229  0.7530
#>  -1.4028  0.4686
#>  -7.5096 -5.6284
#>  -6.3599 -2.7944
#>  -0.9325 -1.2938
#> 
#> (3,2,1,.,.) = 
#>  -0.2065  1.5212
#>  -2.3305 -4.6018
#>  -2.1453 -2.3398
#>   2.9532  1.1517
#>   3.8393 -1.1983
#> 
#> (1,3,1,.,.) = 
#>    6.8507   9.4219
#>   -5.6666   0.4126
#>   -7.1332  -4.1373
#>   -1.1608   0.7581
#>   13.1513  -0.2627
#> 
#> (2,3,1,.,.) = 
#>  -4.3405  0.2608
#>  -0.2952 -5.6971
#>  -0.8747  2.9738
#>   8.1868  8.7141
#>  -3.8800 -1.2806
#> 
#> (3,3,1,.,.) = 
#>   5.0473  9.6755
#>   4.4857 -0.6033
#>   0.7566 -2.0486
#>   3.5538  4.0945
#>  -0.2940 -1.9694
#> 
#> (1,1,2,.,.) = 
#>    3.9388  -1.1465
#>   -0.3026  -2.6466
#>   -0.7977  12.2984
#>    2.4261   5.0481
#>    6.7975  -2.1742
#> 
#> (2,1,2,.,.) = 
#>    3.2900  14.0475
#>    2.9867   2.5135
#>   -1.1376  -0.6015
#>   -1.1814   1.7620
#>    5.8803  12.4532
#> 
#> (3,1,2,.,.) = 
#>  -5.5849 -2.6361
#>   8.0348 -2.2055
#>  -8.2316  1.4563
#>  -3.3230  7.7920
#>  -0.8435 -3.6043
#> 
#> (1,2,2,.,.) = 
#>  -1.3839  0.8536
#>  -3.2929  7.8096
#>   1.5254  2.0544
#>   2.3404  7.7046
#>  -12.4655  3.2411
#> 
#> (2,2,2,.,.) = 
#>    5.4451  10.0953
#>   -2.6363   3.0641
#>   -2.3369   4.3897
#>   -2.3413   0.6266
#>   -6.8191  -5.7607
#> 
#> (3,2,2,.,.) = 
#>    1.2618   3.5082
#>   -4.1416  -4.6586
#>    6.1284  -3.1341
#>   -1.3885  15.2949
#>    6.1557  -2.5673
#> 
#> (1,3,2,.,.) = 
#>  -3.9109 -1.5086
#>   2.9654 -10.9574
#>  -0.1151 -4.0673
#>   8.1934  2.1100
#>   6.1332  1.3104
#> 
#> (2,3,2,.,.) = 
#>  -4.2925  3.9313
#>   6.2656 -6.0244
#>  -8.1150  0.6590
#>   0.7245 -2.5684
#>  -1.3972  1.1685
#> 
#> (3,3,2,.,.) = 
#>   -0.2824   2.0667
#>   11.9034   5.7070
#>   -3.8982  -3.6739
#>    1.9132   2.1163
#>    2.4630  -3.7794
#> 
#> (1,1,3,.,.) = 
#>   4.6526  0.9170
#>   1.6482 -12.8590
#>  -2.1791  2.4462
#>  -1.9041 -0.3237
#>   1.1884 -0.2258
#> 
#> (2,1,3,.,.) = 
#>  -4.4049  0.5988
#>  -4.6665  3.1538
#>   7.8941 -0.0525
#>  -6.5042 -0.1487
#>  -2.5395 -2.0977
#> 
#> (3,1,3,.,.) = 
#>  -0.9548 -4.4487
#>   1.3097 -1.1855
#>   0.0212  3.1801
#>  -2.4676 -9.9708
#>   1.6666 -1.9782
#> 
#> (1,2,3,.,.) = 
#>  -8.5735  6.2634
#>   0.5791  2.8244
#>   2.8510  3.4368
#>  -9.3702 -7.4383
#>   0.1878 -4.1675
#> 
#> (2,2,3,.,.) = 
#>   5.8068 -5.0793
#>   0.8266  2.8248
#>  -10.9557 -1.8647
#>  -3.4674 -7.5204
#>   6.0584 -2.2346
#> 
#> (3,2,3,.,.) = 
#>  -4.2345  1.1481
#>   3.6334  0.8997
#>   7.8529 -3.3739
#>  -0.1228 -6.1366
#>  -0.2876 -0.2984
#> 
#> (1,3,3,.,.) = 
#>   2.1432  5.4002
#>  -6.2483  6.3073
#>   3.1395  9.5883
#>   0.5764 -8.3862
#>   3.5378 -0.6809
#> 
#> (2,3,3,.,.) = 
#>  -2.6364 -2.7025
#>   3.7605 -1.3593
#>   0.3509 -1.8551
#>   0.9081  0.9931
#>   9.6701 -0.6958
#> 
#> (3,3,3,.,.) = 
#>  -4.1449  0.5863
#>  -0.9235 -0.6618
#>  -1.6974 -6.7323
#>   7.0254 -2.3356
#>   1.6898 -3.1400
#> 
#> (1,1,4,.,.) = 
#>  -8.9306 -1.1815
#>  -7.8157 -1.2337
#>   4.2679  2.6224
#>   1.8943 -0.5614
#>   2.5526  0.2732
#> 
#> (2,1,4,.,.) = 
#>   4.3523  0.8043
#>   7.5977  5.5882
#>   0.5486 -2.0111
#>   2.3967  0.4078
#>   5.9060  3.1707
#> 
#> (3,1,4,.,.) = 
#>    1.9517   4.6254
#>   13.1102  -4.5583
#>    1.3060  -5.0397
#>    1.4903   7.7875
#>    2.2769   3.8794
#> 
#> (1,2,4,.,.) = 
#>  -4.5924 -3.2699
#>   8.6970 -0.9823
#>  -2.0957 -10.8813
#>   4.6751 -1.4468
#>  -3.0970 -0.1340
#> 
#> (2,2,4,.,.) = 
#>    1.2514  10.5112
#>    5.3556  -1.4188
#>    3.7613  -0.9527
#>    3.1761   7.8824
#>   -5.0364  -6.4794
#> 
#> (3,2,4,.,.) = 
#>   3.4028 -8.5820
#>  -8.3409 -8.2477
#>  -4.4316 -0.6064
#>   8.0881 -3.8598
#>   6.6616  5.6185
#> 
#> (1,3,4,.,.) = 
#>   0.1445  9.6576
#>   9.5280  3.5922
#>  -0.2394 -5.6020
#>   3.0283  0.4772
#>   2.0265  0.4077
#> 
#> (2,3,4,.,.) = 
#>  -2.6665  0.2748
#>   6.9891  7.4010
#>   4.9652  0.5769
#>   0.3230  3.5149
#>   0.9944  0.9283
#> 
#> (3,3,4,.,.) = 
#>   1.3250 -7.7891
#>  -1.8191 -0.8352
#>   9.1654  3.0695
#>  -5.6724 -0.9505
#>   2.3011  1.8357
#> 
#> (1,1,5,.,.) = 
#>  -7.5601  0.3512
#>   4.1087  1.5842
#>  -2.1942  1.0796
#>  -4.6762  8.5453
#>  -4.2846  0.6583
#> 
#> (2,1,5,.,.) = 
#>   2.5116 -1.1514
#>  -0.9341 -1.3971
#>   0.3646  4.2067
#>  -6.7390  5.0239
#>  -4.6801  2.8120
#> 
#> (3,1,5,.,.) = 
#>    3.7632   0.6102
#>    9.9537   0.8394
#>    0.9448  11.6503
#>    6.9679   2.5092
#>   -7.0604   6.9822
#> 
#> (1,2,5,.,.) = 
#>  -3.3827 -0.4820
#>   0.4445  1.6409
#>  -2.0360 -6.0606
#>  -1.7642 -2.4694
#>   4.2822  7.3040
#> 
#> (2,2,5,.,.) = 
#>   -7.2087  -4.4370
#>   -7.7412  -3.5782
#>   10.2745   7.7468
#>    5.7964   4.5404
#>   -2.9601  -3.8603
#> 
#> (3,2,5,.,.) = 
#>  -6.3566  0.3210
#>   3.0848 -2.2654
#>  -3.2406  0.3190
#>   3.2543  0.0345
#>   1.4277 -4.7216
#> 
#> (1,3,5,.,.) = 
#>  -1.8994 -5.8790
#>   6.9526  9.1929
#>  -0.2060 -0.3320
#>  -5.6400 -4.5748
#>  -3.1138 -1.1250
#> 
#> (2,3,5,.,.) = 
#>   3.1275 -4.3998
#>  -0.4920 -2.9925
#>   1.5847 -6.0685
#>   6.1380  8.2625
#>   0.9799 -6.7389
#> 
#> (3,3,5,.,.) = 
#>   0.7573  1.0305
#>   5.9436 -3.7160
#>  -4.3051  9.4436
#>  -6.7196  1.7264
#>  -8.3305 -7.8613
#> [ CPUFloatType{3,3,5,5,2} ]