Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   3.1313 -9.8605
#>  -10.3096  2.1680
#>   3.2145 -6.7972
#>   8.7178  3.7104
#>   5.3215 -4.5966
#> 
#> (2,1,1,.,.) = 
#>   9.2720  5.3321
#>  -8.0478 -0.9212
#>   4.1212 -0.7513
#>  -0.9211  0.9266
#>  -2.3135 -6.2297
#> 
#> (3,1,1,.,.) = 
#>   4.6218 -2.2312
#>   1.1316  7.8227
#>  -5.4314  4.1894
#>  -5.5477  4.2450
#>  -1.1475  3.4141
#> 
#> (1,2,1,.,.) = 
#>   0.3477 -6.7415
#>   5.4746 -7.5406
#>   1.4646  8.0806
#>  -5.0323 -4.8105
#>   2.4644  8.6131
#> 
#> (2,2,1,.,.) = 
#>   0.6013  9.0992
#>  -13.8289  1.5001
#>  -1.7058 -2.6759
#>   1.3130  5.3859
#>  -7.9293  0.4182
#> 
#> (3,2,1,.,.) = 
#>   5.7364  0.0069
#>   7.5618 -1.6176
#>   3.5128  0.6870
#>   8.2763 -4.0141
#>   4.8888 -5.9535
#> 
#> (1,3,1,.,.) = 
#>   0.0695 -1.3644
#>  -6.0572 -1.9281
#>   2.7976  0.8486
#>  -0.5939 -1.6322
#>   6.5400  8.5459
#> 
#> (2,3,1,.,.) = 
#>  -1.7194  1.4407
#>  -9.6387  6.8653
#>  -2.2596 -1.7017
#>  -0.8391  1.5318
#>  -12.3838  2.0444
#> 
#> (3,3,1,.,.) = 
#>  -3.9757  3.6824
#>  -2.8439  0.5400
#>  -2.8672 -2.4192
#>  -1.9062  2.9075
#>  -2.6272  3.1332
#> 
#> (1,1,2,.,.) = 
#>  -2.6318  5.6955
#>  -3.7870  3.1155
#>   2.1097 -1.0187
#>   0.9036  3.8683
#>  -5.6981  4.2856
#> 
#> (2,1,2,.,.) = 
#>   0.0517 -1.7081
#>   6.9452  3.9824
#>  -6.9958  0.0062
#>  -1.0829  0.5091
#>   0.3249 -3.7108
#> 
#> (3,1,2,.,.) = 
#>  -6.6012 -0.3184
#>  -0.6555  0.6184
#>   0.2725  1.6760
#>   0.4104 -4.0774
#>   2.1442  1.8027
#> 
#> (1,2,2,.,.) = 
#>  -9.2297 -1.1995
#>  -2.6978  4.8814
#>  -6.2417 -4.5938
#>  -2.1380  3.4569
#>   4.9235 -4.2610
#> 
#> (2,2,2,.,.) = 
#>  -5.2243  2.7667
#>  -5.7656  1.0392
#>   0.2448 -6.1037
#>  -7.4469  3.4854
#>   4.8527 -7.0722
#> 
#> (3,2,2,.,.) = 
#>   0.8960  1.5284
#>  -11.4224 -4.4786
#>   2.7783  0.0037
#>   4.4307  3.3031
#>   6.3213 -2.1885
#> 
#> (1,3,2,.,.) = 
#>   8.4583 -2.9497
#>   5.7431  5.4509
#>  -3.8699  2.2671
#>   1.4293 -0.1808
#>  -4.4824 -4.0259
#> 
#> (2,3,2,.,.) = 
#>   0.7586 -2.4183
#>  -1.8652  0.5019
#>   2.0659 -2.1227
#>   4.0533  4.8024
#>  -0.1899  6.6782
#> 
#> (3,3,2,.,.) = 
#>   5.8906  1.8902
#>   9.8738  5.2500
#>   0.5376  4.6917
#>   2.3351 -9.7277
#>   8.0170 -1.2417
#> 
#> (1,1,3,.,.) = 
#>  -10.5518   4.0799
#>   -2.6955  -3.3366
#>   -0.2598   4.0954
#>   -2.7626 -10.8544
#>    3.9937 -12.4364
#> 
#> (2,1,3,.,.) = 
#>   5.8579  4.6544
#>  -6.6684 -3.5691
#>   1.2585 -4.9803
#>  -4.8447  6.5629
#>  -1.9205  2.8372
#> 
#> (3,1,3,.,.) = 
#>   4.0545  2.9366
#>   2.3735  4.8401
#>  -1.7738  1.0509
#>   5.6470 -0.0787
#>  -3.3277  0.5576
#> 
#> (1,2,3,.,.) = 
#>  -0.4127  1.6410
#>   3.6304 -7.4660
#>  -0.1989  0.7276
#>  -8.5324 -1.9093
#>  -14.2932 -3.5760
#> 
#> (2,2,3,.,.) = 
#>   0.7226  5.3374
#>   4.3678  3.5911
#>  -2.1502  6.0900
#>  -3.5676 -9.3835
#>  -3.6494  3.6237
#> 
#> (3,2,3,.,.) = 
#>  -0.0436 -2.3618
#>   1.2322  4.5520
#>  -2.1788 -1.4626
#>   7.4986 -4.7801
#>  -5.5096 -2.8700
#> 
#> (1,3,3,.,.) = 
#>   9.4728  3.7826
#>   6.8633 -6.1030
#>   3.9408  2.9145
#>  -1.1122 -4.2984
#>  -3.0380  4.0166
#> 
#> (2,3,3,.,.) = 
#>   3.8730 -0.7638
#>  -7.2086 -4.3827
#>  -3.4633 -4.2412
#>   5.0375 -0.6136
#>  -4.8122  2.4227
#> 
#> (3,3,3,.,.) = 
#>  -5.1703  2.3465
#>  -2.0737 -11.6889
#>   3.9749  5.9183
#>   3.8392  6.4124
#>  -5.2912 -11.4683
#> 
#> (1,1,4,.,.) = 
#>   2.1684  3.9075
#>  -0.6470  2.3699
#>  -0.4929  0.9232
#>   1.5382  4.8238
#>  -2.4474 -7.7811
#> 
#> (2,1,4,.,.) = 
#>  -2.9355 -1.2338
#>   3.5241 -2.1772
#>   1.8140 -1.9655
#>  -7.3500  0.8795
#>  -6.6091  4.4249
#> 
#> (3,1,4,.,.) = 
#>  -1.5246  8.3792
#>  -9.1824  0.5157
#>  -1.3786  1.4402
#>   4.9558  4.5637
#>  -4.6715  0.8785
#> 
#> (1,2,4,.,.) = 
#>  -4.4492 -3.6085
#>   0.5610 -14.0800
#>   1.4954  5.2705
#>  -6.6287  5.8874
#>  -9.9483  3.1810
#> 
#> (2,2,4,.,.) = 
#>   8.2788 -7.6773
#>   0.1814  5.3330
#>  -0.6776 -0.5223
#>  -2.9245 -8.3487
#>  -0.1334  6.9234
#> 
#> (3,2,4,.,.) = 
#>   0.8615  0.6808
#>  -4.9122 -1.5107
#>   8.2416  3.7771
#>  -0.9657  8.4560
#>  -4.2917  0.9654
#> 
#> (1,3,4,.,.) = 
#>  -4.8084  8.6699
#>  -6.8301  1.9341
#>  -7.5040 -6.6801
#>   5.0753  4.8550
#>   1.8877 -0.2583
#> 
#> (2,3,4,.,.) = 
#>   5.3107 -6.2747
#>  -8.6486  3.4076
#>  -9.7015  1.6776
#>   3.1263 -4.5515
#>   6.5524 -9.9763
#> 
#> (3,3,4,.,.) = 
#>   5.9483 -4.8006
#>  -3.4181  1.0562
#>   8.0127  1.0627
#>   0.1616  2.9524
#>  -1.6360  3.4783
#> 
#> (1,1,5,.,.) = 
#>  -7.7747 -1.0423
#>   2.7201 -0.1651
#>  -1.5976  7.9751
#>   0.4361  5.9732
#>  -2.6047  5.3838
#> 
#> (2,1,5,.,.) = 
#>  -5.3324  5.3911
#>  -7.7645  7.5128
#>   4.8856  2.8171
#>   0.9058  1.7261
#>   2.0247  1.6467
#> 
#> (3,1,5,.,.) = 
#>  -7.5241  1.3054
#>  -9.7162  0.0914
#>   0.6919 -4.4879
#>   0.3817  1.0066
#>   3.6435  5.6845
#> 
#> (1,2,5,.,.) = 
#>    3.2176  -1.6466
#>    4.3241   0.2808
#>   -8.6285   1.6695
#>   -0.4556  -0.5652
#>   -1.4924  10.5845
#> 
#> (2,2,5,.,.) = 
#>  -2.9821  4.6437
#>   1.6212 -3.1428
#>   5.7139 -3.7465
#>  -4.9240  2.4470
#>   5.8293 -2.2502
#> 
#> (3,2,5,.,.) = 
#>   -6.2675   5.2733
#>   -4.5419   2.3285
#>   -1.8674  -0.3982
#>   10.6997  -2.2377
#>    5.6421  -4.1957
#> 
#> (1,3,5,.,.) = 
#>   -7.4039  -3.2396
#>   -4.3349   8.2202
#>    0.1274  -3.5953
#>    0.4219  12.2426
#>   -0.3533   1.0049
#> 
#> (2,3,5,.,.) = 
#>   4.2582  1.0757
#>  -1.7420  1.7918
#>   5.3326  2.4086
#>   1.4504  1.8404
#>   1.3450  7.4969
#> 
#> (3,3,5,.,.) = 
#>  -1.2558  4.7930
#>   4.6823 -5.9583
#>   3.2091  7.2075
#>  -5.8617  6.2166
#>  -2.2951 -1.4857
#> [ CPUFloatType{3,3,5,5,2} ]