Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -1.6962 -3.4542
#>  -0.7952 -2.5022
#>   0.0987 -0.0058
#>   6.8231 -10.1436
#>   3.3296 -10.5816
#> 
#> (2,1,1,.,.) = 
#>  -0.4902  0.0684
#>  -0.0308  0.4312
#>  -3.2327 -10.6928
#>  -1.3012 -5.3116
#>   0.4377 -0.2639
#> 
#> (3,1,1,.,.) = 
#>   -0.8716  -1.9470
#>   -0.1020   1.9353
#>   -6.4348   8.3739
#>   -6.3331  -7.6162
#>   11.8366   9.6752
#> 
#> (1,2,1,.,.) = 
#>   5.6720  0.3650
#>   0.5395 -12.3023
#>   3.1537  3.5083
#>  -10.1358  4.4341
#>  -0.1473  2.3738
#> 
#> (2,2,1,.,.) = 
#>  -0.6929 -9.3803
#>  -0.2215 -9.6819
#>  -1.8675 -4.7426
#>   2.9084  0.3499
#>  -7.2623 -4.0950
#> 
#> (3,2,1,.,.) = 
#>    8.1640  -1.0326
#>   11.0811  -3.7786
#>   -3.9841  -0.2296
#>   11.3769  -2.6495
#>   -6.7664   0.9640
#> 
#> (1,3,1,.,.) = 
#>  -1.3345  3.3505
#>   3.5090 -6.8265
#>   5.3048 -0.8922
#>   5.2860 -2.1915
#>  -1.7596  2.9975
#> 
#> (2,3,1,.,.) = 
#>   0.1494 -0.5145
#>   2.5879 -2.9781
#>  -1.0731 -1.5184
#>  -7.6257 -4.4408
#>   6.9313 -3.1019
#> 
#> (3,3,1,.,.) = 
#>   -4.1527   3.1210
#>    6.9083  12.1086
#>    3.6755   0.8371
#>    4.4916   0.0852
#>    4.0593  -9.0836
#> 
#> (1,1,2,.,.) = 
#>   -1.0867  -0.4890
#>   -2.8422   0.9328
#>   -0.1033  -7.8772
#>    5.0231  -2.2558
#>    3.8789  13.7393
#> 
#> (2,1,2,.,.) = 
#>   2.0430 -0.7134
#>  -6.8803  0.9331
#>  -2.4837  7.3275
#>   5.6572 -9.6809
#>  -2.5495 -2.6296
#> 
#> (3,1,2,.,.) = 
#>   -0.0198   2.3903
#>   -5.5670  10.7820
#>   -1.4896  -2.9134
#>    6.0363   0.3522
#>   -3.1026   7.4027
#> 
#> (1,2,2,.,.) = 
#>  -2.3753  2.4770
#>  -0.8758 -1.5236
#>   6.6391  8.6829
#>   2.3298 -0.6496
#>   3.6663 -4.9696
#> 
#> (2,2,2,.,.) = 
#>   0.8511 -0.5608
#>   1.1894 -5.7132
#>  -0.2387 -3.9971
#>   5.2410 -0.9267
#>  -0.5184  3.0854
#> 
#> (3,2,2,.,.) = 
#>   5.0615 -6.3343
#>  -0.3437  0.0043
#>   0.1753 -13.5302
#>  -11.6716 -4.7549
#>   3.4365  7.5874
#> 
#> (1,3,2,.,.) = 
#>  -1.5880 -2.7878
#>  -1.2802  4.0090
#>  -6.7114  4.7259
#>  -6.6025 -6.8834
#>  -1.9569  6.3559
#> 
#> (2,3,2,.,.) = 
#>   12.2084   2.6373
#>   -8.6397   2.3682
#>    6.5777  -5.1989
#>   -4.0177  -1.8488
#>    2.4542   1.6998
#> 
#> (3,3,2,.,.) = 
#>  -9.0306 -7.6726
#>  -7.7648  5.1996
#>   5.4362  2.2256
#>  -8.8082 -3.4485
#>   2.7953  0.6649
#> 
#> (1,1,3,.,.) = 
#>   -9.4067  12.1079
#>   -4.1126  -3.9528
#>    6.2015  -0.6046
#>   -0.4705  -4.6952
#>   -0.9099  -7.3664
#> 
#> (2,1,3,.,.) = 
#>   1.4467  5.4300
#>  -4.2444 -2.5528
#>  -11.5972  1.8908
#>  -2.0234  3.4979
#>  -2.7054 -0.4948
#> 
#> (3,1,3,.,.) = 
#>   10.9774   4.6580
#>    9.1793  -7.2629
#>    0.9691  -3.0364
#>   -5.1467  -6.2568
#>   -6.8424  -9.4436
#> 
#> (1,2,3,.,.) = 
#>    4.3155   3.9607
#>   -0.5848   2.2799
#>   13.0065   2.9072
#>    0.2868   1.0368
#>    5.5768  -1.1319
#> 
#> (2,2,3,.,.) = 
#>   5.2934 -9.2930
#>   1.0938 -5.0165
#>  -2.7798  7.2327
#>  -8.7858 -2.4293
#>   3.6949  1.4275
#> 
#> (3,2,3,.,.) = 
#>   10.8348   0.0685
#>    2.6510   4.3905
#>    2.6918  -0.6476
#>    1.9115   9.5566
#>    3.8965  -0.6921
#> 
#> (1,3,3,.,.) = 
#>   0.6865  2.9094
#>   1.3834 -5.1981
#>  -5.1423 -6.0691
#>  -11.2013  0.7945
#>  -6.1333  6.1114
#> 
#> (2,3,3,.,.) = 
#>   14.2625  -2.0797
#>    3.6537  -4.6294
#>   -6.0808  -1.9147
#>    2.9437   0.6999
#>   -4.6081   3.7851
#> 
#> (3,3,3,.,.) = 
#>    4.4804  -0.6059
#>   -4.4778  14.6827
#>   -2.6075  -4.2118
#>   -2.3800   7.4738
#>   -1.5509 -10.2913
#> 
#> (1,1,4,.,.) = 
#>   -7.4110  -0.1482
#>    5.5098   1.8293
#>    1.6077  10.4674
#>   -0.3607   0.2868
#>   -3.9079  -2.8878
#> 
#> (2,1,4,.,.) = 
#>  -1.2408 -3.7968
#>   6.5772  1.8024
#>   4.6775  2.4024
#>   0.9316 -0.8771
#>   3.3420  2.7616
#> 
#> (3,1,4,.,.) = 
#>   3.9931  5.8777
#>   7.2040  5.8381
#>  -2.7331  4.3002
#>   3.9715 -5.2154
#>   6.6797  6.0997
#> 
#> (1,2,4,.,.) = 
#>   5.2253 -4.5041
#>   7.1054 -1.1807
#>  -4.3870  7.0245
#>   5.6176 -5.6169
#>   0.3780  2.0058
#> 
#> (2,2,4,.,.) = 
#>  -2.5076 -3.8963
#>  -4.3360  7.4158
#>   6.9247 -1.3237
#>  -2.5888 -0.5838
#>   1.6443  3.2806
#> 
#> (3,2,4,.,.) = 
#>  -0.3123  9.7594
#>  -3.7251  2.1463
#>  -0.6195  4.5479
#>  -3.8757  6.8917
#>   5.0918 -2.9155
#> 
#> (1,3,4,.,.) = 
#>  -8.3692  5.7570
#>  -5.8882  0.1429
#>   3.9224  2.4171
#>   1.2777 -3.3230
#>  -0.2644 -0.6388
#> 
#> (2,3,4,.,.) = 
#>  -8.1421  9.1425
#>   7.9176  8.1158
#>  -0.2026 -3.0834
#>   1.4524  5.9319
#>  -0.3793 -7.2135
#> 
#> (3,3,4,.,.) = 
#>   4.1356 -5.5723
#>  -10.1359  2.7481
#>   1.8989 -0.8787
#>   6.0310  3.2193
#>   4.6702  3.7887
#> 
#> (1,1,5,.,.) = 
#>   6.1916  8.0509
#>   1.9841  9.5275
#>   4.8754  1.9779
#>  -0.6044  0.2845
#>   1.7711  0.0705
#> 
#> (2,1,5,.,.) = 
#>  -3.4861  5.7137
#>  -1.1791 -2.1926
#>   2.0402 -0.3557
#>  -8.7027  0.8367
#>  -4.2881  4.8409
#> 
#> (3,1,5,.,.) = 
#>   1.8226  3.9811
#>  -7.1241 -2.5656
#>  -2.6629  2.3072
#>  -2.4540  5.8007
#>   6.0518  0.8114
#> 
#> (1,2,5,.,.) = 
#>  -6.1476 -1.1838
#>  -4.0037 -0.3265
#>  -3.1195 -3.9595
#>  -4.2239  8.5656
#>  -6.2595 -1.3226
#> 
#> (2,2,5,.,.) = 
#>   10.1292  -6.7998
#>    2.0695  -0.7383
#>  -11.1134   0.7807
#>   -4.8453  -7.2546
#>    0.9136  -9.0256
#> 
#> (3,2,5,.,.) = 
#>   4.2582  1.8877
#>  -5.1349  5.6003
#>   8.7440 -4.3195
#>  -4.3591 -1.2873
#>  -3.2239  0.1494
#> 
#> (1,3,5,.,.) = 
#>  -6.4826  8.2273
#>  -5.2854  6.3710
#>   6.5042  1.5749
#>   4.7526 -3.5676
#>   3.8294 -1.2092
#> 
#> (2,3,5,.,.) = 
#>   0.0449 -5.0681
#>   1.9734  6.2563
#>  -2.5074 -0.3598
#>   1.3133 -5.8335
#>  -0.0424  4.6196
#> 
#> (3,3,5,.,.) = 
#>  -9.3698  7.4339
#>  -0.9072 -6.3306
#>  -3.0791  2.0743
#>   2.0073  2.9950
#>   0.6323  4.9583
#> [ CPUFloatType{3,3,5,5,2} ]