Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   5.2565 -0.3029
#>  -0.6821 -2.1175
#>  -6.8497 -10.8036
#>  -2.2489 -3.8719
#>   4.1756 -0.1043
#> 
#> (2,1,1,.,.) = 
#>   0.2447 -2.7001
#>  -2.8992  2.8825
#>  -2.8851  5.4594
#>   7.5520 -0.5057
#>  -4.8428 -2.1175
#> 
#> (3,1,1,.,.) = 
#>   1.8429 -3.2028
#>  -3.9775  4.0331
#>  -5.5218  5.5038
#>  -4.9261  0.9397
#>   6.3133 -0.0635
#> 
#> (1,2,1,.,.) = 
#>   6.2860  3.3746
#>   4.7605  4.1946
#>   7.5071 -2.3675
#>  -3.4917  3.4110
#>  -5.7536 -3.8601
#> 
#> (2,2,1,.,.) = 
#>  -9.8905  5.6895
#>   6.7449 -5.6321
#>   2.8033 -2.9654
#>  -2.8708 -0.7715
#>  -2.3771 -3.1176
#> 
#> (3,2,1,.,.) = 
#>   0.3143  2.0604
#>  -1.9341 -4.7068
#>   0.3085  8.9640
#>  -5.2404 -3.5509
#>   0.0159  4.7030
#> 
#> (1,3,1,.,.) = 
#>  -2.1215 -6.0591
#>   6.1025 -3.8500
#>  -1.4213 -3.0988
#>   3.4021 -6.4833
#>  -5.8417 -4.7911
#> 
#> (2,3,1,.,.) = 
#>  -0.1188 -14.7816
#>  -4.6644  5.2122
#>   4.9538  4.9538
#>  -2.8110 -1.3097
#>  -3.3044  1.5473
#> 
#> (3,3,1,.,.) = 
#>  -0.6389  4.0754
#>  -11.2018 -5.7073
#>  -2.1633  2.7801
#>   6.0700 -6.6842
#>  -2.3912 -3.0569
#> 
#> (1,1,2,.,.) = 
#>  -10.5105   2.1282
#>   -2.0997  11.9157
#>    1.4722  -5.5625
#>    5.6462  -2.0374
#>    1.4767   2.2251
#> 
#> (2,1,2,.,.) = 
#>   0.0096 -2.9102
#>   3.6335  9.2502
#>   0.5061 -1.6134
#>   9.7831 -2.6517
#>   3.0877  6.8312
#> 
#> (3,1,2,.,.) = 
#>   4.1562 -4.7233
#>   2.3797  2.0454
#>   7.7804 -4.6288
#>  -2.4128 -1.4233
#>  -1.0313 -3.1198
#> 
#> (1,2,2,.,.) = 
#>   6.7669 -0.0455
#>   2.1114 -2.2549
#>  -1.4120  1.2553
#>   2.9533  2.9728
#>   6.1249 -2.8521
#> 
#> (2,2,2,.,.) = 
#>   -4.3828  11.7854
#>   -0.1116  -3.7753
#>    2.3017  -6.1284
#>   11.2306   2.9307
#>    0.1560  -0.4420
#> 
#> (3,2,2,.,.) = 
#>   1.4062  0.8501
#>   1.8258  7.8478
#>   0.3699  3.0950
#>  -1.5037 -4.3380
#>   1.7647 -1.2508
#> 
#> (1,3,2,.,.) = 
#>   -7.5807   8.0701
#>    3.5288   2.7521
#>   -9.7418  -2.1037
#>   12.2007   3.6028
#>   -2.0700  -3.7740
#> 
#> (2,3,2,.,.) = 
#>   0.0379  2.0266
#>  -2.7400  1.2694
#>  -0.9143  2.3790
#>  -0.8446  7.6007
#>   5.2242  9.5142
#> 
#> (3,3,2,.,.) = 
#>  -3.6514 -2.2431
#>  -1.0227  6.4458
#>   3.2464  7.7217
#>  -5.4897  2.6260
#>   5.0966  9.6334
#> 
#> (1,1,3,.,.) = 
#>   -1.0183  -2.8717
#>    0.9216  -1.6284
#>   -8.4933  -1.5368
#>   -5.0030   3.1093
#>   14.9966  -0.6063
#> 
#> (2,1,3,.,.) = 
#>   2.3152  2.2751
#>   5.7075 -0.6842
#>   3.0035  5.4418
#>   1.5556 -2.7109
#>  -4.6568 -12.8880
#> 
#> (3,1,3,.,.) = 
#>  -6.4298  2.5440
#>   1.1418 -3.7239
#>   3.3503  4.5844
#>   1.9698  4.3330
#>  -4.5669  8.4318
#> 
#> (1,2,3,.,.) = 
#>   5.8210  0.7944
#>   2.1538 -3.7296
#>   0.1265 -0.2783
#>   0.6175 -6.3963
#>  -2.0670  4.1047
#> 
#> (2,2,3,.,.) = 
#>   5.3032  4.2134
#>   3.0370  1.7872
#>  -7.6013 -5.3927
#>  -13.0801 -2.3079
#>   8.8999  1.6620
#> 
#> (3,2,3,.,.) = 
#>    3.2883  -4.6429
#>   -3.3802  -2.1221
#>    8.7355  13.1391
#>  -11.7869   2.7838
#>   -1.4592   9.2074
#> 
#> (1,3,3,.,.) = 
#>    1.6456  -4.3790
#>   12.0469   2.4409
#>   -3.8696   0.0188
#>   -3.0616  -3.4768
#>   -9.2044   2.4234
#> 
#> (2,3,3,.,.) = 
#>  -5.7803  0.1416
#>  -9.8864  1.7156
#>   8.4600 -8.6559
#>  -5.3656 -5.2888
#>   7.8387 -3.2869
#> 
#> (3,3,3,.,.) = 
#>  -11.5530   4.1438
#>    4.2577  -5.1448
#>    2.6909  -0.4840
#>    5.7019  -5.7892
#>    2.4189  -3.8008
#> 
#> (1,1,4,.,.) = 
#>    1.4097  -1.5266
#>   -0.2412  -4.9809
#>    1.1906  10.2381
#>   -8.3132  -0.9525
#>    2.8459  -8.1539
#> 
#> (2,1,4,.,.) = 
#>   4.3877 -0.2630
#>   0.5366  7.3806
#>  -4.3775 -4.3823
#>   0.0100  4.0068
#>  -7.5885 -5.9020
#> 
#> (3,1,4,.,.) = 
#>  -7.1684  1.0255
#>   9.6226  6.4362
#>   3.8737 -1.0817
#>  -5.2677  3.0831
#>   6.7056  5.9524
#> 
#> (1,2,4,.,.) = 
#>   6.0886  2.2001
#>   8.7135  7.1844
#>   3.2798 -2.1896
#>  -6.5844 -9.2933
#>  -11.3162 -5.9777
#> 
#> (2,2,4,.,.) = 
#>   4.5606  7.3556
#>   6.0191  1.0967
#>   6.9747 -3.1470
#>  -3.9212  6.6950
#>   2.4901 -4.8745
#> 
#> (3,2,4,.,.) = 
#>    9.4163  10.8031
#>   -4.5978   2.8351
#>   -4.7531   2.9555
#>    1.1174  -6.4291
#>    3.5645   2.4025
#> 
#> (1,3,4,.,.) = 
#>    1.6065  -3.1038
#>    7.9758   2.8094
#>    4.4514  -1.7497
#>   11.1724  -0.0277
#>   -5.2624   6.8979
#> 
#> (2,3,4,.,.) = 
#>   0.3699  6.6338
#>   3.4127 -4.1566
#>  -3.4062  0.7571
#>   7.1137 -2.0926
#>   5.7544 -0.5246
#> 
#> (3,3,4,.,.) = 
#>  -8.8679 -7.1283
#>  -3.0199 -4.8116
#>   0.4743 -4.5047
#>   0.6117 -5.9285
#>  -1.2429  2.9561
#> 
#> (1,1,5,.,.) = 
#>   1.3935  1.2684
#>   1.3837 -3.5616
#>  -2.5587  1.7876
#>   5.5343 -2.2771
#>   2.3489  8.8043
#> 
#> (2,1,5,.,.) = 
#>   1.2608 -0.8054
#>   3.1210  0.8428
#>  -2.9011  4.6624
#>   9.3002  4.3657
#>  -10.3119 -0.3428
#> 
#> (3,1,5,.,.) = 
#>    2.0201  -4.8893
#>   -0.7075   2.8383
#>   -3.8337  -4.8597
#>    4.0306  11.7154
#>    3.3349  -7.3732
#> 
#> (1,2,5,.,.) = 
#>    3.8804   4.4815
#>    0.2211   0.7793
#>   -5.2275  -9.3150
#>   11.7411   4.7682
#>   -0.3359  -2.6551
#> 
#> (2,2,5,.,.) = 
#>  -13.3292  -1.9545
#>   -6.2237  -5.1345
#>    0.6485  -6.2660
#>   -2.3264   3.2271
#>    1.2950   2.2891
#> 
#> (3,2,5,.,.) = 
#>   13.2209  -3.3536
#>  -11.5461  -3.3151
#>   -4.6231   6.9884
#>   12.3169  -0.3304
#>    6.3639  -1.4538
#> 
#> (1,3,5,.,.) = 
#>   0.1033  4.5924
#>  -3.5473 -6.1031
#>  -6.1402  7.6774
#>  -5.4413 -3.7406
#>   2.1764 -15.5501
#> 
#> (2,3,5,.,.) = 
#>   1.4605  1.6028
#>  -3.0898 -11.6767
#>   0.9676  2.2029
#>  -1.6791 -4.9093
#>  -6.6829 -6.5797
#> 
#> (3,3,5,.,.) = 
#>   7.8933 -6.2159
#>  -5.7172 -0.6115
#>  -12.5314  0.2934
#>  -4.3556  2.7290
#>  -1.4360 -5.1483
#> [ CPUFloatType{3,3,5,5,2} ]