Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   -2.5704  -5.8613
#>    9.7329   5.0637
#>   -4.4534  -1.5995
#>    0.4917  -9.6655
#>   -2.7677  12.0905
#> 
#> (2,1,1,.,.) = 
#>   7.9672 -5.9509
#>  -9.0449  4.7169
#>  -1.3165 -3.3204
#>  -0.0149  3.5702
#>   1.3611  3.0640
#> 
#> (3,1,1,.,.) = 
#>  -2.0034  3.4105
#>  -0.2052 -0.6714
#>  -4.2016 -1.9943
#>  -2.8579 -3.8702
#>  -1.8748 -12.2351
#> 
#> (1,2,1,.,.) = 
#>   -6.3257  10.7173
#>    0.4960   9.5347
#>   -3.5141  -0.2427
#>   -3.5820   3.1083
#>    4.8724  -6.9736
#> 
#> (2,2,1,.,.) = 
#>  -9.0339 -6.9923
#>  -0.7574 -6.2759
#>   0.4354 -0.6875
#>  -2.7777  3.5049
#>  -8.8034 -1.9537
#> 
#> (3,2,1,.,.) = 
#>   10.8086  -1.0241
#>   -0.0473   2.2052
#>    6.8273   8.3817
#>    0.1159  -1.5401
#>  -10.3550   2.5985
#> 
#> (1,3,1,.,.) = 
#>   0.6258  7.0307
#>   2.0462  2.4711
#>  -9.5088 -3.7160
#>   6.3572 -9.1390
#>  -3.1569  4.5157
#> 
#> (2,3,1,.,.) = 
#>   1.1513  8.9356
#>   3.3761 -5.4875
#>  -6.9382  8.5093
#>  -2.8848  8.1710
#>   3.1195  3.8438
#> 
#> (3,3,1,.,.) = 
#>   3.3233 -2.6549
#>  -7.6482  2.9781
#>  -4.1913  4.2908
#>   3.8928  3.4448
#>   0.8116 -3.1928
#> 
#> (1,1,2,.,.) = 
#>  -3.6436 -6.4709
#>  -5.9115 -3.0063
#>  -14.3921 -5.7905
#>   0.3431  0.8146
#>  -8.3974  2.2652
#> 
#> (2,1,2,.,.) = 
#>  -0.7531  8.5370
#>   2.5300  3.7942
#>   7.2304 -3.8258
#>  -0.3597  4.5209
#>  -3.3373  0.3587
#> 
#> (3,1,2,.,.) = 
#>    6.1610  -2.1310
#>   -6.7562   0.3269
#>   10.0736  -3.5496
#>    0.2831   0.6828
#>   -5.0459  -4.8099
#> 
#> (1,2,2,.,.) = 
#>  -0.9677  1.2249
#>   1.3771  3.7743
#>   8.2313  3.9640
#>  -3.5030  0.6186
#>  -4.8648 -9.2732
#> 
#> (2,2,2,.,.) = 
#>  -0.0632  2.2425
#>  -3.8493 -3.2353
#>  -5.1040 -7.1735
#>  -2.0483 -4.9333
#>  -6.1488 -4.9197
#> 
#> (3,2,2,.,.) = 
#>  -1.4241 -0.1339
#>  -2.7441  5.8619
#>  -2.5273 -9.8496
#>   3.6603 -3.6857
#>   7.2411 -1.2157
#> 
#> (1,3,2,.,.) = 
#>  -2.7657 -5.6524
#>   1.7908 -0.5746
#>  -4.4008 -4.0581
#>  -0.0733 -5.6839
#>  -1.0954  0.3588
#> 
#> (2,3,2,.,.) = 
#>   6.7612  3.4868
#>  -9.0730  6.0775
#>   2.6386 -1.7402
#>  -0.5997  4.0525
#>   4.9740 -1.3152
#> 
#> (3,3,2,.,.) = 
#>   -2.6236  12.7247
#>    3.5977  -9.1352
#>    4.4415  -3.8805
#>   -1.3956  -5.7210
#>    7.6131   7.5054
#> 
#> (1,1,3,.,.) = 
#>  -4.9616  4.0101
#>  -6.7311  6.8353
#>   4.9294  6.8750
#>   5.9799 -2.3855
#>  -6.5970  7.9822
#> 
#> (2,1,3,.,.) = 
#>  -5.1580  4.4679
#>  -0.4515  4.4353
#>  -0.0034 -5.0947
#>  -0.9458 -1.2590
#>  -6.1743 -3.4467
#> 
#> (3,1,3,.,.) = 
#>  -1.8041 -0.8772
#>  -4.8284 -9.8031
#>  -1.9332 -1.1796
#>   1.3967  6.1436
#>   4.8694  0.0297
#> 
#> (1,2,3,.,.) = 
#>  -7.8238  3.7580
#>  -10.9184 -0.1273
#>   0.3683 -9.7621
#>   6.0604  2.5673
#>   5.9271 -2.3655
#> 
#> (2,2,3,.,.) = 
#>   9.0328  1.3859
#>  -2.1383 -2.7027
#>   4.1710  2.5903
#>  -14.1838 -12.7209
#>   1.0791 -2.8330
#> 
#> (3,2,3,.,.) = 
#>  -1.5577 -0.3798
#>   5.8248  0.6326
#>  -0.5243 -5.6191
#>   8.5503 -5.7810
#>  -0.7448 -1.0660
#> 
#> (1,3,3,.,.) = 
#>  -10.2093  -0.8431
#>    1.2751  -0.7438
#>    0.8054  -8.4704
#>    0.4126  -6.4414
#>    1.8150   4.8704
#> 
#> (2,3,3,.,.) = 
#>  -1.4239  0.4621
#>  -7.0298 -5.1411
#>  -6.3219  3.1561
#>  -0.1074 -3.3115
#>  -0.5788  2.5550
#> 
#> (3,3,3,.,.) = 
#>  -1.7250 -7.6331
#>  -0.5839  7.5710
#>   5.2698 -6.0835
#>  -6.8630  8.7461
#>   6.0272 -5.0477
#> 
#> (1,1,4,.,.) = 
#>   3.1563 -8.2087
#>  -0.4798  1.4050
#>   2.3114 -1.3052
#>  -8.1051 -0.4101
#>  -1.8853 -6.4493
#> 
#> (2,1,4,.,.) = 
#>  -8.7502 -5.4236
#>   8.2196 -3.8891
#>   5.0380 -5.8345
#>   3.1575 -1.3969
#>   6.6744 -7.7787
#> 
#> (3,1,4,.,.) = 
#>   4.0591 -10.0361
#>  -2.0139 -7.9356
#>  -5.9973 -0.6294
#>  -1.6370 -8.5138
#>  -8.1041  2.3220
#> 
#> (1,2,4,.,.) = 
#>  -4.6247  6.5595
#>  -0.6802 -0.8212
#>   2.0011 -1.6926
#>  -14.0686 -6.0118
#>  -4.1130 -13.4013
#> 
#> (2,2,4,.,.) = 
#>   5.0885 -2.3117
#>   1.3145 -3.6656
#>   1.8482 -3.8941
#>  -1.1120  1.0845
#>  -2.1830  4.8663
#> 
#> (3,2,4,.,.) = 
#>   6.7512 -5.2380
#>  -8.2154  2.0986
#>  -7.5876 -4.4728
#>  -1.1435  4.4991
#>  -0.8720  7.4189
#> 
#> (1,3,4,.,.) = 
#>  -1.4116  2.6286
#>   0.1626  3.9850
#>  -4.3065 -4.6265
#>   6.5033  7.5841
#>   0.9435  0.0940
#> 
#> (2,3,4,.,.) = 
#>  -2.2072  4.0650
#>   2.3067 -12.0160
#>   1.5145 -5.5874
#>   4.7959 -0.1937
#>  -3.5550 -1.1818
#> 
#> (3,3,4,.,.) = 
#>  -5.8603  9.8264
#>   3.8106  7.3590
#>   3.1265  2.2502
#>   4.3523  3.4797
#>  -1.7289 -1.9784
#> 
#> (1,1,5,.,.) = 
#>   3.9074  2.1805
#>  -1.3780  2.4452
#>   1.3188  4.8317
#>  -0.1935  2.5207
#>  -7.7019 -4.0398
#> 
#> (2,1,5,.,.) = 
#>   -6.8080   5.6648
#>   -1.7645  -3.7259
#>    4.5347  13.1468
#>    2.9718   4.6961
#>   -5.8218 -11.6491
#> 
#> (3,1,5,.,.) = 
#>    4.1117   2.0414
#>   11.0846   5.0880
#>    3.7267   9.0983
#>    2.2419   2.8912
#>   -1.2558   4.5919
#> 
#> (1,2,5,.,.) = 
#>  -7.0666  1.3932
#>   2.3235  0.3886
#>  -5.6317 -1.7817
#>   2.5743 -3.0918
#>   3.0073  6.5572
#> 
#> (2,2,5,.,.) = 
#>   2.0025  5.1072
#>   0.1018  5.5344
#>   1.0373 -3.7842
#>  -2.0846 -1.5185
#>   6.1686 -2.6016
#> 
#> (3,2,5,.,.) = 
#>    8.4788  -4.0466
#>    4.5526  -5.6987
#>    1.0018   5.3283
#>   -7.0217   8.8239
#>   -4.9047  13.7641
#> 
#> (1,3,5,.,.) = 
#>   4.8692  6.6829
#>  -1.5247 -3.7774
#>  -4.3590  3.2335
#>   3.6588  3.4268
#>  -7.5275 -5.0806
#> 
#> (2,3,5,.,.) = 
#>  -4.8556 -4.7867
#>  -3.1112 -3.8936
#>  -4.2433 -7.8036
#>   0.4058 -0.5670
#>  -8.1098 -3.1983
#> 
#> (3,3,5,.,.) = 
#>  -0.4095 -0.8859
#>  -8.3800 -2.8860
#>  -4.2355  4.3788
#>  -3.2049  3.6902
#>  -7.9717 -2.6632
#> [ CPUFloatType{3,3,5,5,2} ]