Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   2.2181 -3.6087
#>   0.0113  1.3808
#>  -0.1854  4.5949
#>  -6.1453 -3.8144
#>  -5.1800  6.5782
#> 
#> (2,1,1,.,.) = 
#>   2.7116  7.8878
#>   3.2861  2.1512
#>   2.5506  9.0303
#>  -2.9844 -2.7980
#>   3.9154 -7.1962
#> 
#> (3,1,1,.,.) = 
#>  -1.2778 -0.9678
#>   1.8696 -10.2282
#>   2.3266  6.4140
#>   0.0730 -3.9276
#>   0.9301  5.7810
#> 
#> (1,2,1,.,.) = 
#>   0.5298 -0.1887
#>   5.6808  6.3901
#>  -9.1005 -5.5779
#>  -1.0241  0.5789
#>   3.3332 -2.5860
#> 
#> (2,2,1,.,.) = 
#>   0.1891 -0.2577
#>   7.1277  0.6140
#>  -4.0296 -3.6401
#>   1.1902  1.1072
#>  -5.1448 -2.3048
#> 
#> (3,2,1,.,.) = 
#>  -5.0168 -15.8156
#>   4.0136  3.7344
#>   4.1300  3.4690
#>  -5.2821 -0.6709
#>   2.8660 -10.4429
#> 
#> (1,3,1,.,.) = 
#>  -7.5661 -3.7733
#>  -0.7096 -4.2894
#>   0.0146 -4.3052
#>  -0.8685 -2.5549
#>   6.5069  2.8876
#> 
#> (2,3,1,.,.) = 
#>   1.6833 -1.7213
#>   1.1421  2.5430
#>   0.4037  1.4479
#>  -10.2408 -2.4317
#>   6.9068  6.3696
#> 
#> (3,3,1,.,.) = 
#>  -2.5971 -1.9060
#>   3.6991  9.4223
#>   1.1192 -2.7355
#>  -2.2543 -8.8037
#>   4.9513  3.6753
#> 
#> (1,1,2,.,.) = 
#>  -0.5273 -3.4446
#>  -0.6338  0.2753
#>  -0.4054 -0.5658
#>  -5.5014  4.7899
#>  -1.7076  7.6864
#> 
#> (2,1,2,.,.) = 
#>   1.6238  7.9109
#>  -7.3226 -2.6711
#>  -5.6437  2.7441
#>   2.7658 -10.1167
#>  -2.7735  3.4429
#> 
#> (3,1,2,.,.) = 
#>   1.4727 -6.7120
#>  -5.4876 -3.8210
#>  -4.3417  2.2038
#>  -5.2623 -1.8701
#>   6.8679 -5.4115
#> 
#> (1,2,2,.,.) = 
#>   4.4748 -2.4309
#>   2.6630 -2.9024
#>   0.7383 -3.2835
#>  -5.6552 -3.5923
#>  -5.7216 -5.6111
#> 
#> (2,2,2,.,.) = 
#>   6.1835 -1.6863
#>   0.8188 -2.1755
#>  -6.4694  2.0183
#>   1.8691 -5.8208
#>   7.8929 -1.6662
#> 
#> (3,2,2,.,.) = 
#>  -8.1404  1.3486
#>   6.6992  2.5445
#>  -0.4756 -3.8945
#>  -4.3996 -3.9032
#>  -1.0258 -9.1991
#> 
#> (1,3,2,.,.) = 
#>  -2.0212 -2.8682
#>  -1.9560 -0.3367
#>   6.8737 -9.0035
#>   5.3203  7.9776
#>  -1.9017 -9.3781
#> 
#> (2,3,2,.,.) = 
#>  -3.6437  3.8308
#>  -0.0243 -4.6539
#>   5.3651  1.9439
#>   5.3332  1.1659
#>  -1.9107  5.7477
#> 
#> (3,3,2,.,.) = 
#>   3.0743 -4.3475
#>  -3.9885  4.4310
#>   3.1168 -3.7379
#>  -0.9777  1.6836
#>  -1.8899 -1.6526
#> 
#> (1,1,3,.,.) = 
#>  -0.4495  6.2575
#>   5.4448 -1.1216
#>  -3.5234  0.4425
#>   2.7767 -8.2167
#>   1.7032 -4.7431
#> 
#> (2,1,3,.,.) = 
#>   2.3535 -0.1658
#>   6.7803 -1.6728
#>  -2.2855 -7.0310
#>  -8.2345 -0.5778
#>  -2.6603 -8.5015
#> 
#> (3,1,3,.,.) = 
#>   7.3565 -2.9576
#>   6.1145 -0.4352
#>   4.6102 -2.5915
#>   0.4428  1.9922
#>   3.1031 -0.7349
#> 
#> (1,2,3,.,.) = 
#>    7.2106   3.0821
#>   -2.1889  11.4357
#>    2.3955  -6.1094
#>    2.3331   8.8195
#>    0.7127  10.1521
#> 
#> (2,2,3,.,.) = 
#>  -11.5655  -1.2808
#>   -1.8745  -7.0850
#>   -3.9229  -4.9844
#>   -8.8079  -6.7464
#>   -3.1964  -2.8171
#> 
#> (3,2,3,.,.) = 
#>  -7.8765  0.0652
#>   8.8932 -4.3722
#>  -8.9840 -2.5730
#>  -5.1562 -0.8209
#>   1.1591 -4.4994
#> 
#> (1,3,3,.,.) = 
#>   0.9468 -1.6492
#>  -0.4519 -0.5035
#>  -0.6606 -1.7917
#>   3.3117 -12.7631
#>   3.4842  7.1919
#> 
#> (2,3,3,.,.) = 
#>  -5.0322 -0.9286
#>   2.5548  1.8643
#>  -12.3881 -4.3812
#>  -1.1542 -3.2228
#>   3.3547  1.3354
#> 
#> (3,3,3,.,.) = 
#>  -6.3544 -3.2524
#>  -1.1894  7.3566
#>   2.2558  7.1179
#>  -3.1388 -5.9018
#>  -0.2834 -13.2298
#> 
#> (1,1,4,.,.) = 
#>   0.6949 -0.6577
#>  -7.6100  5.2301
#>   4.1988  2.4335
#>  -0.5243 -8.1997
#>   4.6027  5.3917
#> 
#> (2,1,4,.,.) = 
#>   5.4664  5.1122
#>   8.0909  2.0055
#>  -2.7225 -5.8999
#>  -6.1474  1.8665
#>  -1.7992  3.6942
#> 
#> (3,1,4,.,.) = 
#>   1.8115  6.5383
#>  -1.5061 -2.4642
#>   0.4804  9.6527
#>  -3.4743 -7.2445
#>  -3.8115 -1.8303
#> 
#> (1,2,4,.,.) = 
#>   2.7304  4.6289
#>  -11.2267  6.7480
#>   1.9692  2.8580
#>   3.9174  8.5896
#>  -6.9679  0.0453
#> 
#> (2,2,4,.,.) = 
#>  -0.7973  4.3619
#>  -0.5600  1.2773
#>  -3.2400  0.7604
#>   2.0199 -9.8674
#>   2.2427  3.8951
#> 
#> (3,2,4,.,.) = 
#>    2.9590  -0.9163
#>    7.6531   2.0026
#>    1.5585  -1.7018
#>    0.8105   0.0111
#>   -3.5447  10.2590
#> 
#> (1,3,4,.,.) = 
#>   6.1976  2.7465
#>   2.1589  0.4278
#>   8.5013 -6.5324
#>   3.5950  0.7595
#>  -8.2765 -1.3274
#> 
#> (2,3,4,.,.) = 
#>  -12.2168   5.4844
#>   -3.2284   4.8702
#>    4.4085  -4.6555
#>   -3.4613   1.0933
#>    3.9349   2.9665
#> 
#> (3,3,4,.,.) = 
#>   3.7690 -5.8018
#>   2.2372  3.9901
#>   1.0633 -1.6187
#>  -3.4853 -4.3549
#>   3.9789  1.0281
#> 
#> (1,1,5,.,.) = 
#>  -1.7984  0.8008
#>   0.9640 -3.7531
#>  -8.8179 -0.5759
#>   4.4565 -3.7279
#>   0.3376 -1.1993
#> 
#> (2,1,5,.,.) = 
#>  -0.3063  2.7483
#>   0.8693 -8.1362
#>   5.0108  5.9786
#>   0.7802  0.5215
#>   0.1877 -4.9495
#> 
#> (3,1,5,.,.) = 
#>  -0.8756  0.2008
#>  -5.6886 -7.4142
#>  -4.6840 -3.5214
#>   5.8797  2.6045
#>   2.8991 -1.4975
#> 
#> (1,2,5,.,.) = 
#>  -4.7874  6.0504
#>  -2.0248 -3.3290
#>  -2.3952  4.7097
#>  -1.1035 -3.5242
#>  -7.1713  2.3949
#> 
#> (2,2,5,.,.) = 
#>   9.4748 -7.9959
#>  -3.2644  3.7036
#>   4.6814  5.2627
#>   4.2403 -0.8772
#>   0.7005  6.5464
#> 
#> (3,2,5,.,.) = 
#>   -1.3897  -1.4575
#>   11.6668  -0.7169
#>   -4.7981  -0.8109
#>    2.3446  -2.6341
#>   -6.4267  -0.4413
#> 
#> (1,3,5,.,.) = 
#>   1.5666 -0.7376
#>  -6.8020  9.2792
#>   8.1158  7.5261
#>  -10.7027 -1.3319
#>   5.0632  3.6372
#> 
#> (2,3,5,.,.) = 
#>   -2.2640   5.9577
#>   -4.8172   9.1967
#>   -8.1027  -5.5496
#>   10.6911  -8.5767
#>   -2.2522   0.9430
#> 
#> (3,3,5,.,.) = 
#>   2.6596 -5.4818
#>  -0.0639  3.8423
#>  -5.6523  4.8304
#>  -3.7947 -2.2320
#>  -0.7334 -5.3776
#> [ CPUFloatType{3,3,5,5,2} ]