Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -3.8906 -0.2521
#>  -0.6301  7.3887
#>  -0.1569 -3.4600
#>   6.8864  4.7697
#>  -4.4483 -4.8385
#> 
#> (2,1,1,.,.) = 
#>    8.1809  -9.1632
#>   13.5965  10.6084
#>    7.5879   1.5757
#>   -2.9251   9.8662
#>   -3.5436  -0.0408
#> 
#> (3,1,1,.,.) = 
#>  -6.8657  0.6179
#>  -2.1867 -0.9080
#>   3.0976  1.3844
#>  -4.6698  4.5027
#>  -4.1710 -1.5937
#> 
#> (1,2,1,.,.) = 
#>  -4.5069 -1.0340
#>  -1.7006  5.2808
#>  -1.1689 -2.7179
#>   8.6645  5.0827
#>  -4.1420  2.3250
#> 
#> (2,2,1,.,.) = 
#>   7.5556 -1.3444
#>   2.5145 -2.9687
#>   0.8374  1.2748
#>   7.4609  1.1490
#>  -2.5103  6.5001
#> 
#> (3,2,1,.,.) = 
#>   -3.1516  10.3080
#>    5.4505   7.7295
#>    1.9037   0.0331
#>   -2.4931   2.9686
#>    3.7238  -2.8382
#> 
#> (1,3,1,.,.) = 
#>  -1.8638  4.4898
#>   6.2693 -3.0226
#>  -3.7461 -0.5782
#>  -4.6917  4.8592
#>  -7.3360  7.1371
#> 
#> (2,3,1,.,.) = 
#>  -6.3858  2.2408
#>  -5.9384 -2.9792
#>   2.7604  1.3434
#>  -7.1528  0.3610
#>  -1.0356  3.4105
#> 
#> (3,3,1,.,.) = 
#>   8.0636  4.3733
#>  -6.8172  4.5569
#>  -8.2686  9.6735
#>  -6.1795  8.9975
#>   3.4867  1.4089
#> 
#> (1,1,2,.,.) = 
#>    4.0458   1.1211
#>    6.7274   0.8219
#>    3.2647  -8.1180
#>   -4.3900   8.7062
#>   -6.9060  12.5835
#> 
#> (2,1,2,.,.) = 
#>  -14.9166  -0.4563
#>   -8.9484   3.7034
#>    7.5586   0.5212
#>   -6.2767  -7.2894
#>   -2.6380   0.9689
#> 
#> (3,1,2,.,.) = 
#>  -5.0481  4.8095
#>  -0.1282 -0.0783
#>  -3.2317 -2.6416
#>  -0.9148  3.6559
#>  -5.0307  3.9920
#> 
#> (1,2,2,.,.) = 
#>  -1.9028  2.8111
#>  -8.5834 -3.5393
#>   6.6530  5.8489
#>  -4.2728  8.3334
#>   0.6024 -1.7717
#> 
#> (2,2,2,.,.) = 
#>   5.4114 -5.5515
#>   4.1821  5.1946
#>   1.0588  8.8242
#>  -0.0587  2.2180
#>  -11.7108  2.6790
#> 
#> (3,2,2,.,.) = 
#>   2.8153  8.5202
#>   0.0949  1.5582
#>  -0.9742 -6.8452
#>  -11.5933  1.6363
#>  -2.0684  5.5530
#> 
#> (1,3,2,.,.) = 
#>   9.8351  0.5795
#>  -1.6079 -6.4322
#>  -2.0161 -7.3104
#>   1.4246 -2.9077
#>   4.1010 -1.4020
#> 
#> (2,3,2,.,.) = 
#>  -4.0648 -2.0372
#>   6.0770  6.2051
#>  -7.0857  3.3361
#>  -3.6015 -0.8916
#>   8.3811  3.9670
#> 
#> (3,3,2,.,.) = 
#>   0.1586  2.9111
#>  -10.1217 -0.9805
#>   5.0078  0.2680
#>   0.1776  2.0732
#>  -1.6860  0.8195
#> 
#> (1,1,3,.,.) = 
#>  -9.8876 -1.3592
#>   3.7233 -8.8646
#>  -0.5858  2.5747
#>  -0.6867 -0.9947
#>  -7.1288 -6.7674
#> 
#> (2,1,3,.,.) = 
#>  -0.2221 -2.7402
#>  -4.0179  0.0893
#>   7.6763 -6.2895
#>   3.8782  6.3572
#>   0.5241 -0.7963
#> 
#> (3,1,3,.,.) = 
#>  -1.7983  1.1807
#>   2.8927  2.6928
#>  -2.2799  0.1756
#>   6.7964 -8.9340
#>   0.0358  1.1838
#> 
#> (1,2,3,.,.) = 
#>   2.4838 -2.3196
#>  -0.8169  0.6886
#>  -5.2591 -5.7622
#>   0.9913  4.7963
#>  -1.9588  5.2229
#> 
#> (2,2,3,.,.) = 
#>  -1.7353  5.1988
#>   2.5770 -3.1907
#>  -1.6854 -3.0060
#>   1.8460 -0.4476
#>   3.2443 -3.0184
#> 
#> (3,2,3,.,.) = 
#>  -3.8917 -4.1796
#>   4.3798 -2.0535
#>  -5.1279  3.5968
#>   9.4164  3.9474
#>  -1.9752  5.4251
#> 
#> (1,3,3,.,.) = 
#>   0.4361 -5.6941
#>   1.8463 -3.5674
#>   9.1693  1.7494
#>  -6.2192  1.9202
#>   4.6439 -1.7834
#> 
#> (2,3,3,.,.) = 
#>  -10.0798  -1.5540
#>    3.2683   6.1536
#>    5.1260   7.6848
#>   -1.8577  -3.3287
#>    6.9663   1.9440
#> 
#> (3,3,3,.,.) = 
#>  -3.6094  3.3084
#>   4.7373 -5.0252
#>  -3.1775  3.2782
#>  -1.7631  0.7487
#>   5.1359 -5.6054
#> 
#> (1,1,4,.,.) = 
#>    6.2109  -5.7290
#>   -3.2070  -3.6565
#>   10.3394   2.8397
#>   -5.9428   3.2633
#>   -2.0450  -2.4186
#> 
#> (2,1,4,.,.) = 
#>   4.7595  4.1469
#>   0.7921  4.8637
#>  -1.0122 -4.4003
#>  -11.7296  1.0495
#>  -3.6038 -9.4859
#> 
#> (3,1,4,.,.) = 
#>   2.5091  0.5782
#>   1.8351 -4.7400
#>  -4.6407 -2.1848
#>  -2.8872 -3.2071
#>   3.2865 -3.3006
#> 
#> (1,2,4,.,.) = 
#>  -0.7310  6.7275
#>   1.1613 -2.5606
#>   3.1072 -3.6625
#>   0.1855  3.8225
#>   0.3006  1.0869
#> 
#> (2,2,4,.,.) = 
#>   3.9528  3.7990
#>   1.7951 -3.5298
#>  -7.1744 -0.6763
#>   2.3892  4.9950
#>   8.5449 -6.1794
#> 
#> (3,2,4,.,.) = 
#>   5.5085  2.2022
#>  -4.3377  3.1459
#>   7.6732 -3.6306
#>   1.8619  7.0566
#>   7.4051 -7.7637
#> 
#> (1,3,4,.,.) = 
#>  -1.2423  1.1820
#>  -0.4600  1.5733
#>  -8.5312 -2.3545
#>   0.4854 -0.7314
#>  -2.3858 -4.7297
#> 
#> (2,3,4,.,.) = 
#>  -3.8325 -0.5313
#>   0.6302  5.9889
#>  -6.4304  8.7342
#>  -5.5349 -1.8498
#>  -2.5480 -0.2350
#> 
#> (3,3,4,.,.) = 
#>  -5.6014 -5.7887
#>  -5.6342 -0.1735
#>  -8.3357 -0.1609
#>  -3.4009  4.3665
#>  -1.2554 -2.9821
#> 
#> (1,1,5,.,.) = 
#>   -4.1281   1.7054
#>   -4.5965   1.2471
#>    8.7289  12.5479
#>   -1.2868   9.5942
#>    1.1662   0.0020
#> 
#> (2,1,5,.,.) = 
#>   4.1019  3.2576
#>   3.4793  0.8775
#>   2.5433  4.6935
#>   5.2767  2.7942
#>  -0.5224 -2.5172
#> 
#> (3,1,5,.,.) = 
#>  -3.0225  0.2177
#>   6.4263  0.8066
#>  -3.0760  1.7617
#>   6.6881 -2.1407
#>  -8.4150  3.4503
#> 
#> (1,2,5,.,.) = 
#>   0.3347 -9.9952
#>   1.0171 -1.1604
#>   1.1945  3.9782
#>  -1.1903  3.4292
#>   3.8634  1.9562
#> 
#> (2,2,5,.,.) = 
#>   6.2246  5.1168
#>   1.6311 -0.1134
#>   3.6844 -3.9802
#>  -4.8940 -4.1660
#>   5.7609  4.0242
#> 
#> (3,2,5,.,.) = 
#>   4.8045 -1.2544
#>   0.6540 -6.0830
#>  -6.5560  0.9239
#>  -7.5844 -4.3400
#>   2.4527 -0.6081
#> 
#> (1,3,5,.,.) = 
#>   0.0012  1.4379
#>   2.3320 -0.5475
#>   3.0302 -1.6159
#>  -8.4547 -7.8288
#>  -0.9196 -3.0474
#> 
#> (2,3,5,.,.) = 
#>  -1.0915 -0.7531
#>   7.7590  2.9822
#>   1.4004  2.6508
#>   5.4837 -4.4701
#>   0.8624  5.0485
#> 
#> (3,3,5,.,.) = 
#>   1.2067  2.4785
#>  -3.7972  2.4715
#>   1.6085  4.8943
#>   1.5530 -0.9489
#>  -0.9469 -2.9451
#> [ CPUFloatType{3,3,5,5,2} ]