Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   5.9457 -5.7623
#>  -3.0038 -2.4526
#>   5.4524 -4.4446
#>   2.5122  1.9783
#>   4.8540 -1.9707
#> 
#> (2,1,1,.,.) = 
#>    7.6549  13.2622
#>    3.4173   3.0710
#>    6.1769   4.3189
#>    1.0669 -10.5495
#>    2.8352   0.4959
#> 
#> (3,1,1,.,.) = 
#>  -0.2804  4.5876
#>   7.1490 -12.3212
#>   7.5544  0.8287
#>  -0.7094  1.0002
#>   2.5208  4.2768
#> 
#> (1,2,1,.,.) = 
#>   2.7234 -5.3082
#>   3.7939 -11.0197
#>   5.3994  1.3591
#>  -1.4436  2.0106
#>   2.6655 -1.5761
#> 
#> (2,2,1,.,.) = 
#>   1.8700 -4.4250
#>  -8.8579 -3.5870
#>  -0.5300 -0.6330
#>   1.5176  2.6680
#>  -2.2074 -0.8852
#> 
#> (3,2,1,.,.) = 
#>   1.0934 -5.6647
#>  -11.6363  0.6530
#>   0.7920 -4.0179
#>   5.4493  9.2608
#>  -4.8500 -3.8214
#> 
#> (1,3,1,.,.) = 
#>  -1.5190  1.6512
#>  -1.6669  1.9646
#>   2.0424 -2.4020
#>  -2.0992  1.8495
#>  -1.3677  0.5517
#> 
#> (2,3,1,.,.) = 
#>   4.6088 -3.8753
#>  -2.8643 -4.1078
#>  -3.9336 -7.3554
#>   5.2588 -0.8386
#>  -2.3922 -3.0822
#> 
#> (3,3,1,.,.) = 
#>   0.4719  5.7005
#>   0.8266  9.3641
#>  -0.1996  3.6757
#>   7.1329  0.0394
#>  -9.5401  3.6129
#> 
#> (1,1,2,.,.) = 
#>   8.3021 -3.3321
#>   0.4380  1.4571
#>  -4.7327 -3.3804
#>   0.7278  0.4310
#>   4.0153 -10.5926
#> 
#> (2,1,2,.,.) = 
#>   4.0200 -3.6563
#>   1.1721  1.7797
#>   1.0850 -3.4905
#>   1.5211 -0.2699
#>  -1.7279 -1.1411
#> 
#> (3,1,2,.,.) = 
#>  -6.0165 -2.8432
#>  -0.8185 -5.4846
#>   6.5715 -7.4987
#>  -3.9824  0.4338
#>  -4.8709  8.4825
#> 
#> (1,2,2,.,.) = 
#>  -0.3621  2.0840
#>  -5.9376 -1.0465
#>   6.4402 -4.2729
#>   2.6819  4.5068
#>   0.4491 -6.6801
#> 
#> (2,2,2,.,.) = 
#>   2.8812  2.3719
#>  -8.1261 -2.8762
#>   3.9802 -0.4111
#>   1.7739  1.2671
#>  -2.6532 -1.5837
#> 
#> (3,2,2,.,.) = 
#>   10.8449   3.0398
#>    7.6982   3.1588
#>    1.5310  -4.1114
#>    0.2420   4.6567
#>    4.3459   6.5797
#> 
#> (1,3,2,.,.) = 
#>  -0.3305  5.3938
#>   1.1883 -3.2313
#>   1.9374 -2.4877
#>   2.9624 -2.9658
#>  -0.9000  9.6757
#> 
#> (2,3,2,.,.) = 
#>    2.5430  -2.4294
#>    3.9353  -4.1799
#>    7.2429   0.5444
#>    2.0317  11.6103
#>   -5.4516   6.4797
#> 
#> (3,3,2,.,.) = 
#>   1.8980  1.6949
#>  -1.4427  8.2434
#>   2.5452 -1.3258
#>  -0.9848  0.3327
#>  -0.3623 -9.2350
#> 
#> (1,1,3,.,.) = 
#>   -8.9455  -2.2452
#>   -0.7950   7.5456
#>   -0.2510   5.1425
#>   -0.2714  -6.2488
#>    0.0447  11.3562
#> 
#> (2,1,3,.,.) = 
#>   -1.7096  11.7633
#>    1.5837  -6.1559
#>    0.9050   0.2399
#>    9.1244   1.2442
#>   -9.8416   2.4915
#> 
#> (3,1,3,.,.) = 
#>   0.1764 -2.4756
#>   6.9783 -4.0726
#>   0.3142  4.1211
#>  -5.6098 -1.6774
#>  -5.9359  7.3309
#> 
#> (1,2,3,.,.) = 
#>   7.3814  3.2944
#>   8.2556 -7.5471
#>   4.0425 -12.8002
#>  -7.0446  1.4986
#>   0.8601 -3.0507
#> 
#> (2,2,3,.,.) = 
#>  -8.0798  3.3643
#>   0.4606 -2.0119
#>  -0.8308 -1.3406
#>   3.5063 -8.9434
#>   0.9292 -4.1458
#> 
#> (3,2,3,.,.) = 
#>  -11.1667  -2.6017
#>   -7.2479   8.7056
#>    2.5051   5.2216
#>   -2.0147 -14.6879
#>   -0.5802   0.3752
#> 
#> (1,3,3,.,.) = 
#>  -4.7402  5.5858
#>   2.5470 -4.2191
#>  -13.7710 -4.5509
#>   1.0000 -3.7204
#>   0.5107 -6.5634
#> 
#> (2,3,3,.,.) = 
#>   0.2136 -4.0145
#>   4.6607 -14.1099
#>  -2.2136  1.4069
#>   3.4629 -12.6847
#>  -4.6106 -4.9560
#> 
#> (3,3,3,.,.) = 
#>   4.8026  1.7287
#>   0.8946  5.9304
#>  -5.8787  1.2693
#>  -2.2353  5.7205
#>  -3.2856 -1.1071
#> 
#> (1,1,4,.,.) = 
#>   5.0217  1.5295
#>  -0.7173  1.9389
#>  -2.7840 -1.9681
#>  -0.3334 -5.7188
#>   0.4493 -1.9525
#> 
#> (2,1,4,.,.) = 
#>   -4.0217   1.0108
#>   -1.7965  10.5716
#>   -0.8064  -3.0868
#>    1.5232  -2.2430
#>    5.8064   0.3777
#> 
#> (3,1,4,.,.) = 
#>   0.6244  3.3884
#>  -0.2719 -3.4515
#>  -6.2940 -4.7500
#>  -4.0819  4.2565
#>  -4.0238  3.4784
#> 
#> (1,2,4,.,.) = 
#>    7.4675  10.5056
#>    5.9253   3.0402
#>   -5.5857  -2.4729
#>    9.2926   0.0407
#>    0.6840  -5.3551
#> 
#> (2,2,4,.,.) = 
#>    2.6629  -8.3272
#>   -6.4911   6.8235
#>   -8.9170   0.2326
#>    2.3658   8.9967
#>    0.0326  11.7039
#> 
#> (3,2,4,.,.) = 
#>  -5.7751 -2.1870
#>   2.7912 -0.2455
#>   0.5684  9.9245
#>  -9.7848  3.9908
#>  -6.2620 -5.2242
#> 
#> (1,3,4,.,.) = 
#>   0.3269 -3.7260
#>  -1.7785  8.4960
#>  -3.5982  5.8164
#>  -7.5270 -4.4663
#>   5.5641  4.0842
#> 
#> (2,3,4,.,.) = 
#>  -8.4397  6.4716
#>  -2.2355  3.4852
#>   0.6337 -7.9787
#>   7.5603 -2.7222
#>  -2.7611  1.3885
#> 
#> (3,3,4,.,.) = 
#>   6.5515  5.6199
#>   1.7911  1.5633
#>   4.6462  3.4555
#>  -3.2063 -0.3074
#>  -4.1960  7.7819
#> 
#> (1,1,5,.,.) = 
#>    5.6066   4.1272
#>   -2.3763  -2.6466
#>   -0.3081  -0.6886
#>   12.4722  -4.1436
#>   -9.7105   3.5103
#> 
#> (2,1,5,.,.) = 
#>   0.5194  0.9374
#>   5.6908 -7.5367
#>  -2.8586  3.2810
#>  -3.3327 -5.4591
#>  -12.5403 -0.7304
#> 
#> (3,1,5,.,.) = 
#>   7.3186 -0.5398
#>  -5.4398  9.5187
#>  -4.9687 -2.0043
#>  -0.9192 -1.7891
#>  -9.3830 -6.5444
#> 
#> (1,2,5,.,.) = 
#>   4.7453 -1.9418
#>   5.6458  7.5857
#>  -5.3955  4.2495
#>   7.1400  1.8908
#>   0.8288 -8.3268
#> 
#> (2,2,5,.,.) = 
#>  -4.9140 -7.2619
#>   5.7802 -4.2236
#>  -4.2840 -7.1960
#>   5.6346 -6.8484
#>   4.6211 -5.1984
#> 
#> (3,2,5,.,.) = 
#>   1.5513 -3.6497
#>   1.0055 -3.5779
#>   6.5766 -8.5655
#>  -0.6301 -1.5478
#>   2.1162  3.5193
#> 
#> (1,3,5,.,.) = 
#>   8.1860 -0.2201
#>   3.7715  3.5339
#>   5.9187  2.5389
#>  -3.6267 -0.7601
#>   0.6459 -3.5511
#> 
#> (2,3,5,.,.) = 
#>   1.2623  0.4328
#>   7.0974  4.1673
#>  -1.8856  0.9383
#>  -0.8504 -3.0870
#>   5.5747  3.2681
#> 
#> (3,3,5,.,.) = 
#>  -3.3368 -0.7776
#>  -1.9238  0.4305
#>  -6.2042 -2.7040
#>  -13.8514  0.3180
#>  -3.2077  3.4025
#> [ CPUFloatType{3,3,5,5,2} ]