Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   0.5645 -3.7521
#>   1.7856 -5.2852
#>  -5.2667 -3.2534
#>   2.0833 -6.8790
#>   1.2305 -9.9604
#> 
#> (2,1,1,.,.) = 
#>    0.4720   2.6429
#>   -0.1394  -2.7411
#>    3.3616   9.5459
#>   -7.3020   5.2591
#>    2.1782  12.2962
#> 
#> (3,1,1,.,.) = 
#>   1.5621 -15.8471
#>  -6.4453 -3.1589
#>   8.4296 -7.8871
#>   3.8198 -0.7798
#>  -6.3351  4.9466
#> 
#> (1,2,1,.,.) = 
#>   2.7194  0.4405
#>   2.2404  5.4768
#>  -9.6153  0.0090
#>   1.3590  1.3922
#>  -0.4145  8.6998
#> 
#> (2,2,1,.,.) = 
#>   -7.7434  -2.3125
#>   11.6662  -4.6395
#>    2.9707   1.8616
#>    5.0800   2.6558
#>    7.0923   9.2961
#> 
#> (3,2,1,.,.) = 
#>  -5.8010  5.1898
#>   1.8949 -2.1924
#>   1.9525 -2.7085
#>   3.7733  4.5271
#>   0.5595 -0.6718
#> 
#> (1,3,1,.,.) = 
#>  -1.5768  2.2396
#>  -1.3257 -0.5147
#>   9.9179 -4.7653
#>   2.9926  0.7526
#>   2.2346  0.2047
#> 
#> (2,3,1,.,.) = 
#>  -5.1926  4.5383
#>  -0.3311  0.2490
#>  -9.5984  3.4337
#>  -8.2616  0.9692
#>   1.3035  8.4503
#> 
#> (3,3,1,.,.) = 
#>  -1.6088  6.1236
#>   3.1875  2.5181
#>  -5.0320  0.3274
#>   0.1213  0.5018
#>  -4.2217  1.7533
#> 
#> (1,1,2,.,.) = 
#>   1.4994 -4.8029
#>   4.4191  0.7953
#>   7.7172  2.5681
#>  -4.4751 -0.0701
#>   1.5622 -8.2084
#> 
#> (2,1,2,.,.) = 
#>   3.0461 -1.9172
#>   4.6920 -0.9425
#>   4.4991  2.1458
#>   0.9396 -2.6322
#>  -2.8034 -0.5349
#> 
#> (3,1,2,.,.) = 
#>   1.2483 -2.6344
#>  -4.9525  9.7465
#>  -6.2801  9.8007
#>   4.7737 -8.8490
#>  -2.4079  0.2178
#> 
#> (1,2,2,.,.) = 
#>  -10.4468   5.1502
#>   -3.8841  -5.7667
#>   -6.1502  -2.8648
#>   -7.0605  -1.0422
#>   -5.5507  -4.6373
#> 
#> (2,2,2,.,.) = 
#>   1.5576  7.0773
#>  -2.3515 -0.1200
#>  -8.3232 -8.0879
#>   0.7406 -5.6142
#>  -3.7018 -3.8973
#> 
#> (3,2,2,.,.) = 
#>   2.3790 -4.1338
#>   3.9337 -3.7162
#>   0.5015  1.2214
#>   2.1993 -0.3852
#>  -1.0852 -8.1298
#> 
#> (1,3,2,.,.) = 
#>   0.0047 -1.8272
#>  -2.6577  3.8114
#>  -0.2088  2.9509
#>   6.4224 -1.1408
#>   0.5636 -4.8275
#> 
#> (2,3,2,.,.) = 
#>  -2.9640 -2.6803
#>  -7.7598  5.7401
#>  -4.4524 -0.3289
#>  -1.0306  5.5677
#>  -5.1605  0.7707
#> 
#> (3,3,2,.,.) = 
#>   2.7057  0.6111
#>  -3.4817 -3.6951
#>  -1.3117  0.0267
#>   2.0281 -3.2586
#>   0.7678  0.4352
#> 
#> (1,1,3,.,.) = 
#>  -6.8015  3.4276
#>  -3.8828  0.0986
#>   8.9868 -0.7066
#>   6.2744  7.9653
#>  -0.1749 -7.9577
#> 
#> (2,1,3,.,.) = 
#>    0.5151   7.3233
#>    2.3619  10.5532
#>    2.7106  11.6500
#>    8.5128   6.9380
#>    9.3485   3.1972
#> 
#> (3,1,3,.,.) = 
#>   11.6138  -2.2556
#>    1.3175  -0.9916
#>    1.7744  -5.2009
#>    8.7241   7.3090
#>    0.6667   1.8952
#> 
#> (1,2,3,.,.) = 
#>   7.1898  2.9207
#>   7.7462 -8.5787
#>   4.9846  0.7246
#>   0.6663 -7.0679
#>   0.5992 -2.4569
#> 
#> (2,2,3,.,.) = 
#>   -1.1675  -2.4542
#>    6.0941   4.5308
#>   11.6183  -9.8796
#>    2.6971   2.0594
#>    2.8118  -0.7865
#> 
#> (3,2,3,.,.) = 
#>   0.7787  1.8668
#>  -1.5913  4.0274
#>   9.7426  0.4943
#>  -3.8119  1.4143
#>   1.6909  7.6891
#> 
#> (1,3,3,.,.) = 
#>   3.0280 -5.1455
#>  -6.6807  0.1627
#>  -5.2630 -1.4613
#>  -0.6610 -4.0006
#>   7.0447  3.6004
#> 
#> (2,3,3,.,.) = 
#>   1.3531 -5.5770
#>  -0.2502  5.3239
#>  -3.5166  0.0305
#>   0.9845 -1.4681
#>   3.8066 -2.4658
#> 
#> (3,3,3,.,.) = 
#>  -0.6769 -3.4697
#>   0.5566  0.6593
#>   3.0076  3.6346
#>  -7.2216  0.4103
#>  -10.2717  1.9541
#> 
#> (1,1,4,.,.) = 
#>  -3.1689 -1.7323
#>  -2.7027  2.8342
#>   0.3203 -2.1283
#>  -0.9763  2.0345
#>  -4.3142  2.1069
#> 
#> (2,1,4,.,.) = 
#>   1.9621  5.8402
#>   0.0847 -0.5024
#>   1.9356 -6.0231
#>   7.1043 -6.7825
#>   9.9500  9.7423
#> 
#> (3,1,4,.,.) = 
#>   0.9818  2.0523
#>  -8.7979 -3.9865
#>   4.1513 -0.6377
#>  -4.0700 -1.8076
#>   4.6677 -0.3003
#> 
#> (1,2,4,.,.) = 
#>   -8.9227  10.1623
#>   -3.5426  -1.3229
#>    7.8018  11.0383
#>    7.0764  -0.6816
#>    0.6050   2.4620
#> 
#> (2,2,4,.,.) = 
#>   3.9932 -3.2612
#>   0.5829  6.5954
#>   4.8772 -4.7568
#>  -0.3538 -2.3409
#>   0.6106  0.8458
#> 
#> (3,2,4,.,.) = 
#>  -3.2333  9.1903
#>  -10.1508  1.0844
#>   0.9549 -12.0900
#>   4.0871 -3.6921
#>   9.3730  3.2482
#> 
#> (1,3,4,.,.) = 
#>  -7.6709 -11.6739
#>   0.4648 -4.5325
#>  -7.4592 -1.2040
#>   0.9900  6.6403
#>  -0.3247  3.3300
#> 
#> (2,3,4,.,.) = 
#>   2.2110 -7.5249
#>   1.1974 -3.0106
#>   5.6430  1.1800
#>   7.8529 -3.3497
#>   8.1917 -0.1094
#> 
#> (3,3,4,.,.) = 
#>    7.5727   1.1278
#>   -3.6182  -2.6864
#>   -9.9617  -0.6491
#>   -4.6703   0.8074
#>   -1.0057  10.8349
#> 
#> (1,1,5,.,.) = 
#>    0.7092   4.6354
#>    2.0325  -1.9544
#>   -2.3895   9.1311
#>   14.5793  -8.6489
#>    1.9892  -1.9732
#> 
#> (2,1,5,.,.) = 
#>   4.3610 -0.0548
#>   5.2205  0.0242
#>  -3.3927 -2.0494
#>   6.9487 -1.2118
#>   0.7540 -2.3577
#> 
#> (3,1,5,.,.) = 
#>  -1.2964 -1.7543
#>  -4.1430 -3.0630
#>   3.7591 -7.4449
#>   1.2391  3.6970
#>  -4.5794  5.1406
#> 
#> (1,2,5,.,.) = 
#>   3.2531 -0.4970
#>  -4.2411  2.2141
#>   0.8552  1.8861
#>   3.5244  1.4057
#>   8.4026  3.9653
#> 
#> (2,2,5,.,.) = 
#>   -2.5299   5.1869
#>   -3.1904   6.4928
#>    4.6746  10.5887
#>    4.5123   3.2002
#>   -0.1181   2.7460
#> 
#> (3,2,5,.,.) = 
#>    7.2557   2.5203
#>    4.4477  -4.3900
#>    3.0780   2.0588
#>   -3.6181  13.4870
#>   -3.2371   0.0413
#> 
#> (1,3,5,.,.) = 
#>   1.7263 -2.6868
#>  -3.6825  1.9451
#>   0.0430  0.3669
#>   3.1257 -5.6873
#>  -1.2930  3.7561
#> 
#> (2,3,5,.,.) = 
#>   4.1664 -4.2160
#>  -2.2694  1.7793
#>  -4.2465 -6.0492
#>   0.1933 -2.4015
#>   5.4280  1.6617
#> 
#> (3,3,5,.,.) = 
#>   -2.7599   2.2101
#>    1.9765  -9.9529
#>   10.3486  -4.6254
#>   -3.4072   0.1741
#>    7.1778   0.5619
#> [ CPUFloatType{3,3,5,5,2} ]