Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   2.3610  3.9453
#>   0.5671 -3.0591
#>  -2.7243  0.5454
#>   7.6076 -5.1644
#>  -3.8999 -3.7912
#> 
#> (2,1,1,.,.) = 
#>  -2.1582 -1.1236
#>   5.8545  0.0479
#>   0.6268  3.0217
#>  -3.1616  6.8735
#>  -2.2099 -0.7198
#> 
#> (3,1,1,.,.) = 
#>   7.9878 -1.5192
#>  -6.3201 -4.6946
#>  -7.4433  5.0549
#>   1.7491  1.7551
#>   2.7375 -0.1162
#> 
#> (1,2,1,.,.) = 
#>  -1.8962 -7.9871
#>   3.0237  2.5546
#>   2.3120  0.8744
#>   6.3798 -0.1539
#>   0.8967  4.1213
#> 
#> (2,2,1,.,.) = 
#>  -2.9779  2.8191
#>  -4.4292  3.3568
#>  -0.5985 -2.8581
#>  -2.6908 -2.7287
#>  -4.7414 -9.6754
#> 
#> (3,2,1,.,.) = 
#>   8.3235  5.6209
#>   8.0391  0.5984
#>  -9.2368  7.2515
#>   3.7964  5.7276
#>  -3.9684 -9.1656
#> 
#> (1,3,1,.,.) = 
#>   2.7281 -8.0165
#>   1.5274  0.2512
#>   3.3755  5.1237
#>  -4.1819 -3.4454
#>   2.2885 -8.3418
#> 
#> (2,3,1,.,.) = 
#>   3.3283 -5.8537
#>   0.7416 -10.9206
#>   3.2558  1.6522
#>  -5.4411  6.4739
#>   4.6460  3.1668
#> 
#> (3,3,1,.,.) = 
#>   3.7263  4.6015
#>  -3.9702 -16.2927
#>  -3.3092 -4.1192
#>  -2.3486 -1.5239
#>  -5.7296  4.3484
#> 
#> (1,1,2,.,.) = 
#>   5.5350  0.2906
#>  -6.0963  2.0256
#>  -5.3645  4.1891
#>   7.1420 -0.8828
#>  -0.9855  4.8041
#> 
#> (2,1,2,.,.) = 
#>   4.7523  0.0748
#>  -1.5044  0.6805
#>  -3.3640 -3.2829
#>  -8.2478 -6.8470
#>  -1.6410 -0.2676
#> 
#> (3,1,2,.,.) = 
#>   2.0711  6.0124
#>   4.1365 -3.5219
#>   5.5411 -2.0816
#>   2.4497 -0.9760
#>   5.9847  5.4580
#> 
#> (1,2,2,.,.) = 
#>   -1.0344   1.2600
#>   -0.1976   3.8196
#>   -0.1577   2.7569
#>    5.1592  11.4006
#>   10.7721   0.5215
#> 
#> (2,2,2,.,.) = 
#>   4.2670 -4.1536
#>  -9.5499  4.2181
#>   5.0027 -8.4970
#>  -2.5931 -3.7855
#>  -3.0292  5.6180
#> 
#> (3,2,2,.,.) = 
#>  -4.4997 -7.3926
#>  -2.0358  2.4209
#>   3.0385 -1.8334
#>  -1.5273 -0.6429
#>   2.3366  1.3413
#> 
#> (1,3,2,.,.) = 
#>   -6.8976  -4.2171
#>   -4.5683  10.5744
#>    0.8663   2.5542
#>    0.5199  10.9520
#>   -5.7630   4.2385
#> 
#> (2,3,2,.,.) = 
#>   7.8743  4.7780
#>  -3.6374 -8.9217
#>  -3.8429 -0.2735
#>   5.1696 -1.2930
#>  -3.9358 -7.9662
#> 
#> (3,3,2,.,.) = 
#>  -0.9477 -0.6589
#>   4.1443 -9.6679
#>  -5.5075  3.0292
#>  -0.9616 -0.6463
#>   2.4271 -8.8636
#> 
#> (1,1,3,.,.) = 
#>   5.1936 -3.6883
#>  -4.2568  0.9304
#>  -8.3124  2.2336
#>  -0.5150 -2.0621
#>   6.0992  7.9199
#> 
#> (2,1,3,.,.) = 
#>  -0.6063 -0.4112
#>  -2.6315 -4.6048
#>  -7.7405  4.4401
#>  -9.0316 -1.4777
#>   7.7684 -6.6259
#> 
#> (3,1,3,.,.) = 
#>   2.9325 -1.7696
#>  -2.0695  2.6817
#>  -0.7644 -8.3379
#>  -1.9193  3.3336
#>  -6.0013 -9.3517
#> 
#> (1,2,3,.,.) = 
#>  -0.4547  6.3539
#>   0.9705 -2.0854
#>   2.1752  5.5018
#>   3.1759 -2.5956
#>   5.7128  1.0729
#> 
#> (2,2,3,.,.) = 
#>   0.2653  7.5732
#>  -0.3078 -2.2559
#>   3.9102  0.4271
#>  -4.6463  0.6886
#>  -6.6236  7.6331
#> 
#> (3,2,3,.,.) = 
#>  -7.3847  1.2643
#>   0.5940 -12.8204
#>  -3.0971 -8.9878
#>   1.5541  0.0897
#>   0.0156  3.3420
#> 
#> (1,3,3,.,.) = 
#>    1.9363   3.4069
#>    2.0242  -1.5555
#>   -1.7471   0.3545
#>   -0.3287   9.0304
#>   11.0957  -5.8803
#> 
#> (2,3,3,.,.) = 
#>  -2.1326  2.2136
#>   0.2978  4.6986
#>   9.0921 -4.2687
#>   0.9820 -10.1727
#>  -0.6007 -8.3729
#> 
#> (3,3,3,.,.) = 
#>   11.4676   4.6704
#>   -2.3310  -6.4118
#>    2.5983  -0.2408
#>    0.2022   8.0914
#>    0.6446   2.0224
#> 
#> (1,1,4,.,.) = 
#>   8.0492  1.4355
#>  -7.1263 -1.2473
#>  -6.7485  1.1679
#>   7.2344 -7.5097
#>  -1.2663 -6.7633
#> 
#> (2,1,4,.,.) = 
#>    2.2239  -5.5334
#>    9.5509  -0.0360
#>   -7.1274  -5.0100
#>   -4.9126  -1.1522
#>   -8.2704  11.1574
#> 
#> (3,1,4,.,.) = 
#>   3.0374  1.8185
#>   8.6972  5.0401
#>  -0.7180  2.1982
#>   1.2598  4.8749
#>  -0.3217  3.3916
#> 
#> (1,2,4,.,.) = 
#>   7.4756  1.3955
#>  -1.3804 -6.4252
#>   1.2669  0.0748
#>   3.5458  2.4000
#>   5.6566  7.6429
#> 
#> (2,2,4,.,.) = 
#>  -5.7257 -10.8743
#>  -2.8111 -5.1875
#>  -2.1043 -1.4108
#>   2.2225  2.8792
#>  -14.2507  3.6613
#> 
#> (3,2,4,.,.) = 
#>    0.5108   5.8060
#>   10.2526   0.0615
#>   -2.2305  -1.3197
#>    7.2798  -0.3733
#>   -1.2591  -3.1435
#> 
#> (1,3,4,.,.) = 
#>    0.5181  -1.6616
#>    0.9827  -3.7321
#>   -0.3522  13.7861
#>    1.4463   7.5154
#>   -2.7381  -9.1113
#> 
#> (2,3,4,.,.) = 
#>  -2.4401  0.1757
#>   1.4442 -6.6350
#>  -4.6430 -6.6419
#>   1.0657  4.3117
#>   1.4418  3.8539
#> 
#> (3,3,4,.,.) = 
#>   2.5294  0.6081
#>  -4.1670  2.0741
#>   4.3382 -0.0704
#>  -5.0666 -1.9383
#>  -1.4389  9.8376
#> 
#> (1,1,5,.,.) = 
#>  -7.5740 -1.1731
#>   4.3251 -5.9972
#>  -10.2923 -5.2612
#>   4.3837 -3.3689
#>   0.8083 -11.0843
#> 
#> (2,1,5,.,.) = 
#>  -6.3317  6.3057
#>   3.5798 -3.4250
#>   0.7338  0.9535
#>  -1.6951  2.7719
#>   3.2537 -5.9143
#> 
#> (3,1,5,.,.) = 
#>  -3.9227  3.5958
#>  -2.5369 -4.2652
#>   5.7757 -1.7687
#>  -0.0812 -4.5679
#>   7.6620  0.6663
#> 
#> (1,2,5,.,.) = 
#>   3.9548  5.4444
#>   7.9551  3.2483
#>   1.9849  0.9298
#>  -3.8232  8.2412
#>   2.9364 -1.7714
#> 
#> (2,2,5,.,.) = 
#>   -1.1893   9.2277
#>   11.2624  -8.5708
#>    4.5798   3.9809
#>   -1.0574  -7.3330
#>   -5.0541  -4.7935
#> 
#> (3,2,5,.,.) = 
#>   4.3221 -5.4182
#>   3.0173 -0.8626
#>   4.2822 -5.5490
#>   7.4381  5.5651
#>  -0.6943  0.4901
#> 
#> (1,3,5,.,.) = 
#>  -7.4593  4.9227
#>  -5.4684  1.3110
#>  -5.7055  4.7758
#>  -3.3464 -1.4652
#>  -3.0396 -10.7163
#> 
#> (2,3,5,.,.) = 
#>   2.1042  0.1855
#>  -5.0917 -0.6214
#>  -11.2179 -5.6920
#>   1.5191  5.8465
#>  -0.2811  0.7535
#> 
#> (3,3,5,.,.) = 
#>  -0.9185  5.7286
#>  -3.8930  0.1554
#>  -0.4865 -2.9486
#>   0.7679 -2.2787
#>  -4.4999  0.8029
#> [ CPUFloatType{3,3,5,5,2} ]