Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -5.1363 -6.2471
#>  -5.7578 -0.2691
#>  -0.1414  6.4603
#>   7.8807 -8.0271
#>  -4.6531 -4.7855
#> 
#> (2,1,1,.,.) = 
#>   3.3499  1.1886
#>  -2.4614  6.6017
#>  -3.4215  6.5250
#>   6.7238  2.9558
#>   3.8009  6.8358
#> 
#> (3,1,1,.,.) = 
#>  -0.2885  3.6832
#>  -4.5814 -5.1812
#>  -0.4005  4.7240
#>   1.1686  0.1533
#>   0.1663  5.9182
#> 
#> (1,2,1,.,.) = 
#>    0.7091 -10.8469
#>    3.0124  -0.5981
#>   -7.4975  -1.1077
#>   12.3712   6.7281
#>   -2.1835   1.3075
#> 
#> (2,2,1,.,.) = 
#>  -4.7222  1.9605
#>  -8.8857  1.8716
#>   1.0201 -3.6548
#>   3.4497 -0.8632
#>   0.1542 -3.4768
#> 
#> (3,2,1,.,.) = 
#>  -5.7290  1.5049
#>   6.1088 -2.8925
#>   0.6484 -1.4136
#>   9.5613  8.2872
#>  -0.7062 -2.0881
#> 
#> (1,3,1,.,.) = 
#>   2.4609 -4.4963
#>   9.1677  4.7484
#>  -4.6699  2.1282
#>  -1.8479 -0.0202
#>   6.1530  4.9667
#> 
#> (2,3,1,.,.) = 
#>   1.9773 -0.3653
#>  -7.8108 -5.8367
#>   5.2183 -6.3179
#>  -3.8610 -8.0250
#>   5.6353 -8.0508
#> 
#> (3,3,1,.,.) = 
#>  -5.6592 -2.6753
#>  -3.8924 -1.1567
#>  -2.8837  2.3638
#>  -0.6025  1.3693
#>  -10.4108  2.0658
#> 
#> (1,1,2,.,.) = 
#>  -6.1716  6.1660
#>   3.1156  9.7578
#>  -2.6333  3.3887
#>   2.1960 -0.1476
#>   0.6985 -3.8399
#> 
#> (2,1,2,.,.) = 
#>  -0.0024 -5.9708
#>  -1.5493 -0.4865
#>  -6.0134 -0.7347
#>   1.0093 -5.5300
#>  -2.0379  1.3807
#> 
#> (3,1,2,.,.) = 
#>  -9.9357  1.4287
#>  -1.3165  2.8557
#>  -1.8631 -3.7059
#>  -4.6952 -1.8857
#>   2.9982 -2.9482
#> 
#> (1,2,2,.,.) = 
#>   4.6386 -5.7340
#>   7.8886  3.0170
#>   1.8104  5.1516
#>   4.0091  4.6702
#>  -0.3846 -2.0937
#> 
#> (2,2,2,.,.) = 
#>  -4.2201  5.5354
#>  -7.4786  8.5812
#>  -5.0470  8.4513
#>   2.8768 -4.6922
#>  -2.7263 -1.5879
#> 
#> (3,2,2,.,.) = 
#>  -8.8902  0.1923
#>  -2.0129 -2.1036
#>   3.3478 -6.2799
#>   0.8804  3.4309
#>   3.4911 -5.6683
#> 
#> (1,3,2,.,.) = 
#>  -9.0231 -4.0766
#>  -1.7201 -4.5021
#>   3.8276 -0.6791
#>   4.5597  7.4248
#>   0.3210  4.3910
#> 
#> (2,3,2,.,.) = 
#>  -1.2471 -1.5252
#>   3.8098  1.3094
#>   2.9030  4.5719
#>  -0.1369 -6.7866
#>  -11.0124 -0.5167
#> 
#> (3,3,2,.,.) = 
#>   6.3124  0.3756
#>  -11.3042 -0.0005
#>  -0.4803  2.4308
#>  -4.7767  0.2334
#>  -7.2626 -3.5327
#> 
#> (1,1,3,.,.) = 
#>   -5.4498  -4.8643
#>    5.8794   3.3249
#>    6.2959   5.5924
#>    4.9714   2.1800
#>   -1.4723  10.8368
#> 
#> (2,1,3,.,.) = 
#>  -0.3681 -5.0101
#>  -4.5379  0.3160
#>  -0.3782  1.7454
#>  -9.7867  8.1684
#>   1.5246  0.2066
#> 
#> (3,1,3,.,.) = 
#>   0.8356 -3.6196
#>  -6.5047  3.6913
#>  -3.4543 -4.2249
#>  -0.6832  5.5101
#>  -0.9657  9.9087
#> 
#> (1,2,3,.,.) = 
#>   4.1547  2.8903
#>  -4.3883 -1.1876
#>  -3.0818 -1.4909
#>  -2.2626  6.1090
#>   8.0513 -1.3081
#> 
#> (2,2,3,.,.) = 
#>   1.4862 -1.5294
#>  -8.2956 -2.0712
#>  -3.9917 -5.4823
#>   0.9484  3.8779
#>  -0.6755  3.3972
#> 
#> (3,2,3,.,.) = 
#>  -1.3397  4.0418
#>   4.0538  3.9649
#>  -1.1837  9.3853
#>   5.3254 -8.1337
#>  -2.6480 -5.8593
#> 
#> (1,3,3,.,.) = 
#>   -7.5586  10.5641
#>   -5.7212   1.2778
#>   -0.8260   7.2446
#>   -0.9826  -0.7035
#>    4.8457   0.9338
#> 
#> (2,3,3,.,.) = 
#>   1.1489  3.1902
#>   2.9344  1.2577
#>   2.1298 -11.2687
#>  -0.9028 -4.1823
#>  -0.6539  1.3399
#> 
#> (3,3,3,.,.) = 
#>    0.5536   2.3170
#>    5.2792   0.7510
#>   11.0515   6.3244
#>    6.2719   4.7460
#>   -0.0081   2.9818
#> 
#> (1,1,4,.,.) = 
#>  -0.6551  4.7873
#>  -3.9171  5.9613
#>   6.9855  5.6985
#>  -6.1622 -10.7397
#>   6.0174 -6.2269
#> 
#> (2,1,4,.,.) = 
#>  -7.4504  8.6362
#>  -0.6029  3.4946
#>   0.9284  3.3605
#>  -4.3733 -2.8496
#>  -2.2908 -2.5390
#> 
#> (3,1,4,.,.) = 
#>  -1.0022 -1.4119
#>  -6.5415 -0.0892
#>   3.3249 -1.5865
#>   2.9872 -2.7307
#>  -6.8238  3.3632
#> 
#> (1,2,4,.,.) = 
#>   1.0997 -2.1007
#>   3.9866  3.1425
#>  -0.8947  5.1982
#>  -4.1228 -1.7720
#>   5.2825  1.1890
#> 
#> (2,2,4,.,.) = 
#>   5.0633 -6.5026
#>   0.7893  3.3881
#>   0.9724  1.0095
#>  -8.3896  5.6066
#>   4.3171  0.1161
#> 
#> (3,2,4,.,.) = 
#>    3.9134   3.0318
#>    6.3665   3.4804
#>    2.9486   3.1322
#>   -3.0936  12.4826
#>   -8.6506   2.3294
#> 
#> (1,3,4,.,.) = 
#>   -6.6942   8.4807
#>    7.7618  -1.4189
#>    1.7696  -2.7723
#>   -5.0685  10.3346
#>    3.2298   4.6008
#> 
#> (2,3,4,.,.) = 
#>  -4.6181 -3.1640
#>   1.5151  2.1336
#>  -6.5989 -3.8983
#>   6.7945 -2.2305
#>  -2.5972 -1.1821
#> 
#> (3,3,4,.,.) = 
#>    1.1747   1.2476
#>   -4.2468  -2.3446
#>    4.5586   8.1712
#>   -1.9218  -7.5042
#>   16.1131 -11.9151
#> 
#> (1,1,5,.,.) = 
#>  -0.9269 -0.1640
#>  -2.9500 -6.3982
#>   3.7427 -2.7211
#>  -9.0953  1.7972
#>   3.4066 -3.7440
#> 
#> (2,1,5,.,.) = 
#>    3.2936   6.9847
#>   -3.4446  -5.2195
#>   11.8412   1.6737
#>   -1.4001  -6.2300
#>   -2.5298  -1.6328
#> 
#> (3,1,5,.,.) = 
#>   2.9450  5.6958
#>  -3.6223  4.6959
#>  -3.6375  1.0640
#>  -2.9060 -8.6688
#>   1.5289 -5.7174
#> 
#> (1,2,5,.,.) = 
#>    2.3051   1.8401
#>    5.2685   2.0985
#>   -0.9271  -5.7369
#>   11.3222  -0.7324
#>   -3.0402   6.5860
#> 
#> (2,2,5,.,.) = 
#>   -2.6340   0.7723
#>    4.8382  11.8955
#>   -2.1951   3.7971
#>    0.9359  -0.2596
#>    7.8694   0.8572
#> 
#> (3,2,5,.,.) = 
#>   -5.4789  -5.0938
#>    1.4230  -2.9676
#>   13.3838  -0.1471
#>   -4.6375  -1.2539
#>    2.5142   6.1859
#> 
#> (1,3,5,.,.) = 
#>  -6.6784 -1.2860
#>  -2.4143 -8.8902
#>  -6.9061 -3.7875
#>   0.2695 -4.8694
#>   2.2679  8.0676
#> 
#> (2,3,5,.,.) = 
#>   4.9634  3.0272
#>   5.6511  1.5021
#>   9.4936  0.0528
#>  -2.6152  4.3783
#>  -0.8155 -0.5300
#> 
#> (3,3,5,.,.) = 
#>  -9.3306 -4.4196
#>  -7.5460  2.4044
#>   0.0350  2.1185
#>   4.8886  5.1963
#>  -4.5266 -1.8827
#> [ CPUFloatType{3,3,5,5,2} ]