Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -8.7294 -0.7641
#>  -2.4599 -4.5370
#>   4.8572 -4.6174
#>   4.9751 -4.4805
#>   5.2779 -1.7157
#> 
#> (2,1,1,.,.) = 
#>   6.8390  5.1270
#>  -1.1831 -2.8975
#>  -8.4560 -6.7523
#>   4.6154  5.7443
#>  -7.3759  0.4670
#> 
#> (3,1,1,.,.) = 
#>   5.8456  3.9577
#>   8.0932  4.1858
#>  -2.5091 -7.3443
#>  -1.8669 -0.0962
#>   1.8266 -1.9322
#> 
#> (1,2,1,.,.) = 
#>   0.3198  2.1284
#>   2.8202  0.6929
#>   7.8397  1.6112
#>  -9.6591 -13.4466
#>   5.6329 -1.4847
#> 
#> (2,2,1,.,.) = 
#>  -3.4481 -5.5138
#>   0.1132  2.8383
#>   0.6987 -5.3231
#>   3.9657  6.4545
#>   0.5429  8.8194
#> 
#> (3,2,1,.,.) = 
#>  -2.7845  4.3661
#>  -0.2051 -4.8747
#>  -3.1496 -1.3473
#>  -1.3175  1.5572
#>  -1.3375  4.4127
#> 
#> (1,3,1,.,.) = 
#>   1.7745 -6.4243
#>   8.8761 -3.3835
#>  -4.4971 -0.8175
#>   7.2352  0.2233
#>   2.8077 -1.8680
#> 
#> (2,3,1,.,.) = 
#>   5.2274  2.3991
#>  -4.7519  2.5354
#>   6.6621  5.5734
#>   6.7713  7.6599
#>   5.5926  4.3271
#> 
#> (3,3,1,.,.) = 
#>  -13.6205   4.7777
#>   -5.8868  -5.2908
#>    7.4566  -1.2078
#>    5.3911   1.7017
#>   -4.1169  -5.2450
#> 
#> (1,1,2,.,.) = 
#>  -5.6191 -0.5454
#>  -2.7425 -0.2278
#>   6.7292  4.7944
#>   9.5222  1.2270
#>   0.4555  2.2785
#> 
#> (2,1,2,.,.) = 
#>    1.2017  -1.5516
#>   -7.1329   3.9662
#>    9.9309  10.5788
#>    0.4492   3.9881
#>   -3.1641   2.5180
#> 
#> (3,1,2,.,.) = 
#>  -2.0766  2.8577
#>   1.2194  4.9328
#>  -1.9045 -1.0366
#>  -2.1718 -0.0786
#>   3.3553 -5.1397
#> 
#> (1,2,2,.,.) = 
#>   6.4527 -3.4474
#>   1.9408 -2.7468
#>  -8.4633 -0.1048
#>  -2.5993 -5.3785
#>  -4.0916 -2.8968
#> 
#> (2,2,2,.,.) = 
#>  -0.4771 -11.1814
#>  -8.4196 -7.7143
#>   3.3186 -8.2276
#>  -2.5150 -7.3816
#>  -1.3463  2.1840
#> 
#> (3,2,2,.,.) = 
#>   1.3837  0.3744
#>   1.8107 -0.0973
#>   5.9407 -4.4143
#>   4.3572  2.3915
#>   0.0193 -6.5915
#> 
#> (1,3,2,.,.) = 
#>  -7.8869 -6.7086
#>  -11.8247 -0.5618
#>  -0.5612  0.6639
#>  -0.0299  5.5108
#>  -5.4819 -4.0859
#> 
#> (2,3,2,.,.) = 
#>  -0.4398  0.3026
#>  -2.0807  7.6766
#>  -2.2631 -1.6765
#>  -13.9067  4.9384
#>  -9.5914 -0.3896
#> 
#> (3,3,2,.,.) = 
#>    6.3580  -8.7962
#>    4.4971   9.0322
#>   -2.3268  -1.7182
#>    7.6241  14.4991
#>    2.9240  -5.5173
#> 
#> (1,1,3,.,.) = 
#>   4.3073  2.8056
#>  -11.8370 -7.1395
#>  -4.3841  5.0465
#>  -2.8432  6.4338
#>   7.3372  5.2510
#> 
#> (2,1,3,.,.) = 
#>   8.7608  1.9279
#>   0.4920  1.8646
#>   3.1675  1.8446
#>   4.7426  3.0904
#>  -7.1657  8.8530
#> 
#> (3,1,3,.,.) = 
#>  -1.4495 -0.0741
#>   0.7354 -2.0999
#>  -8.2208 -0.3902
#>   1.6128 -7.3889
#>   5.4694  6.7901
#> 
#> (1,2,3,.,.) = 
#>   -0.0476  -0.3076
#>   -0.4178  -1.0375
#>   -4.5377  -7.8986
#>   -5.9360   5.3708
#>   -2.6004  10.7311
#> 
#> (2,2,3,.,.) = 
#>    6.7661   3.2812
#>   11.7999   2.5044
#>    1.3158   7.3207
#>   10.5602   0.7147
#>   -6.3084   2.3467
#> 
#> (3,2,3,.,.) = 
#>  -0.8499 -4.7202
#>   6.2820  1.2262
#>   5.3507  4.7785
#>   6.9243 -6.5186
#>   3.2226  6.1511
#> 
#> (1,3,3,.,.) = 
#>  -0.9162  1.1574
#>   2.2674  1.3292
#>   1.9063 -1.5237
#>   3.3462 -0.1398
#>  -0.4710 -5.6876
#> 
#> (2,3,3,.,.) = 
#>  -5.4719  4.7873
#>  -6.0582  4.0897
#>  -1.6625 -7.3766
#>   0.0958 -2.3204
#>   0.2307 -6.4544
#> 
#> (3,3,3,.,.) = 
#>  -3.5457  0.1932
#>   6.0345 -1.8267
#>   2.0689  4.3657
#>   5.0117 -3.5542
#>  -0.9066  2.5641
#> 
#> (1,1,4,.,.) = 
#>   5.0202 -1.0834
#>  -0.5424 -0.6613
#>  -1.2651 -0.3932
#>   0.3913  0.7476
#>   4.4089 -6.1607
#> 
#> (2,1,4,.,.) = 
#>   0.6765 -6.7807
#>  -6.0792 -1.3350
#>   3.1995  0.7972
#>   2.1077 -5.2071
#>   4.9709  4.8206
#> 
#> (3,1,4,.,.) = 
#>    0.3724   4.0351
#>   -3.1418   0.7091
#>   10.5928  12.5481
#>    0.2003  -7.1101
#>    2.0174  -3.3556
#> 
#> (1,2,4,.,.) = 
#>  -9.2603 -0.0808
#>  -6.5668 -1.0255
#>  -11.9513  4.9548
#>  -3.6405  1.3313
#>   0.3744  8.2753
#> 
#> (2,2,4,.,.) = 
#>  -1.7619  0.8238
#>   2.9100 -0.9304
#>   0.0078  5.4219
#>  -1.0267 -3.7274
#>  -7.0861 -2.6026
#> 
#> (3,2,4,.,.) = 
#>   5.7663 -2.1607
#>  -3.5107  1.5649
#>   2.6166 -10.1211
#>  -4.6634 -2.9613
#>  -4.8674 -0.5065
#> 
#> (1,3,4,.,.) = 
#>  -1.7135  1.6185
#>  -3.0617 -1.1789
#>   2.1780 -2.0151
#>   2.6466  0.7300
#>   1.9575  3.5969
#> 
#> (2,3,4,.,.) = 
#>  -2.2881  0.8345
#>  -3.9994 -10.9619
#>  -1.6736  0.7327
#>   2.5959 -12.4208
#>  -4.0915  1.8162
#> 
#> (3,3,4,.,.) = 
#>   10.7079  -0.0646
#>    1.0674  12.5726
#>   -2.0384   3.2183
#>    3.1687  -1.1640
#>    0.3367  -8.0734
#> 
#> (1,1,5,.,.) = 
#>  -5.7559 -2.4152
#>   2.6114 -0.4982
#>   0.9661  4.2835
#>  -0.4130 -1.3694
#>  -3.1117  2.0076
#> 
#> (2,1,5,.,.) = 
#>   -7.8332   4.3916
#>    6.1948  10.0351
#>    0.9997  14.8297
#>   -1.2970  -1.2175
#>    1.3690  -2.1566
#> 
#> (3,1,5,.,.) = 
#>  -5.2123  2.1344
#>  -0.6352  3.7269
#>  -0.7840 -2.9039
#>   0.7943  8.3303
#>  -3.3807  2.9307
#> 
#> (1,2,5,.,.) = 
#>    4.3716   3.7583
#>    6.1037  -1.3243
#>   11.5103   0.4505
#>   -1.0692   2.1549
#>   -3.5912   0.0214
#> 
#> (2,2,5,.,.) = 
#>  -6.5965 -1.5735
#>  -1.7892 -8.9855
#>   9.6146  7.1781
#>   6.5649  0.7543
#>   0.4732  9.4677
#> 
#> (3,2,5,.,.) = 
#>    4.5158  10.3264
#>   -0.9937   3.6709
#>   -6.8074  -3.7242
#>  -11.7420   5.5333
#>   -7.2607  -0.3630
#> 
#> (1,3,5,.,.) = 
#>    1.6205   2.4814
#>   10.3727  -8.2623
#>   10.6452  -0.2015
#>    4.4624  -0.4860
#>   -4.1646  -8.3407
#> 
#> (2,3,5,.,.) = 
#>  -6.7556  6.1886
#>   5.8541 -1.6469
#>   0.2451  9.6876
#>   2.0497 -3.9692
#>  -4.7437  4.2787
#> 
#> (3,3,5,.,.) = 
#>   3.9264 -0.1647
#>   3.0392 -1.4740
#>   2.4292  0.1618
#>   2.5031  5.6737
#>   2.9708  4.5677
#> [ CPUFloatType{3,3,5,5,2} ]