Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>    0.6109  -6.6300
#>   -2.0119   0.8782
#>    1.2171  10.3995
#>    4.0923   0.3201
#>    1.1301   2.9702
#> 
#> (2,1,1,.,.) = 
#>   -8.5475   4.3014
#>    6.0553  -2.3035
#>   10.0191  -1.2773
#>   -6.7297   3.4510
#>    0.1298   2.2474
#> 
#> (3,1,1,.,.) = 
#>  -1.0490  6.8776
#>   1.0058 -9.5403
#>   7.3354 -3.3131
#>   0.5238 -0.0543
#>  -7.7062  4.7551
#> 
#> (1,2,1,.,.) = 
#>  -2.4143 -1.2214
#>  -3.4526 -0.7853
#>  -5.4286 -0.2963
#>   0.5500  0.1300
#>  -4.5563 -7.4163
#> 
#> (2,2,1,.,.) = 
#>  -3.3886 -2.1898
#>  -1.2693 -5.6915
#>   4.4589 -7.6344
#>  -2.2940  3.8206
#>  -1.0637 -3.4148
#> 
#> (3,2,1,.,.) = 
#>  -4.1501 -5.8114
#>  -3.9416 -12.2672
#>   2.4993  3.4425
#>  -7.6225 -2.1909
#>   5.0483  1.3574
#> 
#> (1,3,1,.,.) = 
#>  -0.4686  2.2506
#>  -11.6780 -2.8915
#>   3.8886  1.0199
#>   5.0668 -6.2178
#>   7.1027 -0.0643
#> 
#> (2,3,1,.,.) = 
#>   3.3795 -5.4112
#>   4.3272  0.3013
#>  -2.3593  2.0275
#>   0.9154 -2.5516
#>  -4.7512 -2.4407
#> 
#> (3,3,1,.,.) = 
#>   1.2252 -0.3226
#>  -7.9732 -3.9422
#>  -2.7962  2.4096
#>  -2.0652 -1.7292
#>  -2.3155  0.0714
#> 
#> (1,1,2,.,.) = 
#>   4.2634 -6.6692
#>  -6.4730 -0.4085
#>  -3.9010  2.2564
#>  -3.0164 -5.7608
#>  -3.9707 -3.8483
#> 
#> (2,1,2,.,.) = 
#>  -6.7975  5.0126
#>  -2.2728  0.5513
#>  -1.7987 -1.1275
#>   0.8723 -6.4398
#>   1.9776 -0.2265
#> 
#> (3,1,2,.,.) = 
#>   1.5030  1.9541
#>  -1.3570 -3.3934
#>   4.1756 -3.9052
#>  -1.3425  0.3271
#>   2.5890  0.9050
#> 
#> (1,2,2,.,.) = 
#>    4.5420   2.9545
#>   -1.8777   5.1885
#>    1.2218   4.8843
#>   15.9157  -1.0232
#>   -3.5158   0.5860
#> 
#> (2,2,2,.,.) = 
#>  -1.9559  7.2681
#>   4.8410  6.1783
#>   5.0441 -1.2906
#>  -5.2164  2.3989
#>   0.0226 -1.5920
#> 
#> (3,2,2,.,.) = 
#>  -3.9845 -2.4095
#>   4.3481 -8.4156
#>   1.0706  3.2952
#>   2.3378  5.9071
#>   0.8062  0.3055
#> 
#> (1,3,2,.,.) = 
#>   0.9078 -0.5387
#>   4.5019 -0.6822
#>   1.8375  2.7183
#>   4.1608 -3.1865
#>   3.3276 -13.1290
#> 
#> (2,3,2,.,.) = 
#>   -2.6378  -6.4955
#>   -4.5514 -11.0631
#>    2.7269  -5.9383
#>   -4.3084  10.6824
#>   -5.2207   3.1869
#> 
#> (3,3,2,.,.) = 
#>   -2.9406   6.2574
#>   -5.1372   3.9251
#>    0.8265   3.2530
#>    7.8598   3.0238
#>    6.0810  11.3964
#> 
#> (1,1,3,.,.) = 
#>  -3.9129  9.0110
#>   6.1132 -1.4569
#>  -0.9034  8.3935
#>   3.0205  9.5425
#>   8.7756  5.3063
#> 
#> (2,1,3,.,.) = 
#>  -3.3718  1.5750
#>  -1.4461 -2.9534
#>   3.1302  2.4626
#>  -2.6578 -0.8668
#>   2.6006 -0.7327
#> 
#> (3,1,3,.,.) = 
#>  -0.8044  3.6567
#>  -5.6036 -1.3240
#>  -4.9712 -3.1732
#>  -5.6551  0.0258
#>  -2.5613 -4.4721
#> 
#> (1,2,3,.,.) = 
#>   2.9488  2.9570
#>  -2.9885 -6.3336
#>  -0.6228  3.3750
#>   1.3233  4.8570
#>  -3.1879 -6.9371
#> 
#> (2,2,3,.,.) = 
#>  -3.2673  0.7659
#>   1.3506  3.2617
#>   4.7484 -6.9982
#>  -2.2835  6.5859
#>  -0.8291 -4.3984
#> 
#> (3,2,3,.,.) = 
#>   2.0825 -4.1366
#>   2.5524 -1.4177
#>   2.1856  5.6127
#>  -0.1842 -2.6815
#>  -8.6098 -2.1518
#> 
#> (1,3,3,.,.) = 
#>   7.0352 -2.9699
#>  -5.4013  3.2013
#>   3.5587 -1.4021
#>   6.8363  4.0039
#>  -12.0706  4.8081
#> 
#> (2,3,3,.,.) = 
#>   1.0796 -1.4458
#>  -3.9455 -4.9047
#>   1.2486 -6.6867
#>   3.5650 -0.2018
#>   0.6322  3.5393
#> 
#> (3,3,3,.,.) = 
#>   2.0100  0.8551
#>   4.5467 -6.1240
#>   1.9185  3.9005
#>   0.7520 -6.0013
#>  -2.1033  0.7125
#> 
#> (1,1,4,.,.) = 
#>  -5.9604 -4.4671
#>  -1.5594  5.9540
#>  -1.5685  0.9419
#>  -9.3132  2.4019
#>   1.1665 -0.2206
#> 
#> (2,1,4,.,.) = 
#>   1.6436  0.5837
#>   4.7352  2.3041
#>   7.4473  1.8011
#>   4.9651  3.7908
#>   1.8975 -5.0541
#> 
#> (3,1,4,.,.) = 
#>  -2.7434  4.1847
#>   2.7575  3.5199
#>  -1.6034  0.4595
#>   0.7776 -2.8967
#>   3.8990  1.5340
#> 
#> (1,2,4,.,.) = 
#>   5.8117  0.8692
#>  -1.6844  2.3095
#>  -4.0018  1.2480
#>  -7.1648 -4.2276
#>  -2.3900 -1.7520
#> 
#> (2,2,4,.,.) = 
#>   6.4153 -0.9953
#>  -0.8606  0.0519
#>  -4.4123 -1.3529
#>   2.4645  1.0664
#>   4.7891  1.4182
#> 
#> (3,2,4,.,.) = 
#>   3.2249  4.5148
#>  -2.4042 -5.7177
#>   0.5550  0.7957
#>   0.0238 -4.5168
#>   0.3688 -8.2436
#> 
#> (1,3,4,.,.) = 
#>  -7.7525 -2.5175
#>   4.3951 -6.5409
#>  -1.8363  5.5605
#>   5.4779 -3.4244
#>   0.1532 -0.6372
#> 
#> (2,3,4,.,.) = 
#>    2.3629   3.8926
#>   -6.0552   4.7364
#>    4.5000  -7.9110
#>    1.1075  10.1763
#>   -5.9313   0.9829
#> 
#> (3,3,4,.,.) = 
#>   4.2583 -1.6221
#>   5.2745  0.7960
#>  -0.5403  3.4416
#>   1.0198  0.8096
#>  -1.1176 -2.0387
#> 
#> (1,1,5,.,.) = 
#>  -6.5580  5.3656
#>   3.5976  0.7621
#>   2.1306  1.1717
#>   0.0219  0.7460
#>   0.9981  0.7719
#> 
#> (2,1,5,.,.) = 
#>  -1.1455  5.0021
#>  -6.3161  4.7016
#>   2.1453  6.0336
#>   6.7279 -1.1180
#>   1.3520  9.0918
#> 
#> (3,1,5,.,.) = 
#>  -3.9994  5.2916
#>   1.8232 -0.0237
#>  -1.0618 -0.8827
#>  -1.6342  3.7920
#>  -3.1407  0.2792
#> 
#> (1,2,5,.,.) = 
#>  -3.3195  6.2722
#>   4.4308 -10.0676
#>   6.0728 -6.1752
#>   2.4882 -6.1304
#>   4.9129  8.4485
#> 
#> (2,2,5,.,.) = 
#>  -4.3701 -6.2936
#>  -9.6023  1.2452
#>   2.4811  0.6745
#>  -0.4581 -3.0696
#>   0.7264  3.3454
#> 
#> (3,2,5,.,.) = 
#>  -4.7373 -3.2412
#>   3.0586  7.1875
#>   0.4528  4.3290
#>  -1.3419  8.4725
#>   1.1335  2.3962
#> 
#> (1,3,5,.,.) = 
#>    4.5060  11.2071
#>    1.8982   8.1990
#>   -0.4963   0.7907
#>   -2.5300  -1.6697
#>   -4.1890   3.0598
#> 
#> (2,3,5,.,.) = 
#>   4.6819  8.2868
#>  -2.1526  6.4826
#>   3.3308 -4.7593
#>   5.0425 -1.7795
#>  -0.0070  5.2186
#> 
#> (3,3,5,.,.) = 
#>  -4.3430  3.9149
#>  -7.5206 -3.0119
#>   6.6375 -9.1574
#>  -8.4701  2.8047
#>  -3.5073 -10.0573
#> [ CPUFloatType{3,3,5,5,2} ]