Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -1.3740  6.0452
#>   8.4714 -4.5049
#>  -2.6174 -5.0787
#>   5.3811  8.4647
#>   9.0711  4.5233
#> 
#> (2,1,1,.,.) = 
#>  -1.2374 -2.8372
#>   3.6186  3.2998
#>  -2.5669  3.0506
#>  -0.1556  5.2143
#>   5.5195 -4.8316
#> 
#> (3,1,1,.,.) = 
#>   -4.4583  -8.4876
#>    2.9697   7.0177
#>    0.9681  -5.9261
#>    0.3099   2.0888
#>   10.2729  -0.1894
#> 
#> (1,2,1,.,.) = 
#>  -3.8894 -7.4167
#>  -7.9607  0.2706
#>   0.1114  5.7068
#>  -0.7870 -6.8041
#>   5.1891 -2.7963
#> 
#> (2,2,1,.,.) = 
#>   0.9644 -2.2186
#>   1.1088 -7.0026
#>  -5.1336 -5.5937
#>   2.4775 -8.9563
#>  -7.7373 -0.5736
#> 
#> (3,2,1,.,.) = 
#>   3.5519 -4.9166
#>  -0.0072  5.2711
#>  -7.5589 -12.7636
#>   1.3081  0.9104
#>   5.7309  5.6933
#> 
#> (1,3,1,.,.) = 
#>   2.3925  0.0077
#>  -12.9042  6.4213
#>   3.8302  9.5253
#>   7.4026 -5.7027
#>  -1.8322 -3.9033
#> 
#> (2,3,1,.,.) = 
#>   -2.1954   2.4311
#>    8.4805   8.6964
#>   11.6852   4.9782
#>   -6.1005   4.3583
#>   -5.0296   1.9610
#> 
#> (3,3,1,.,.) = 
#>  -2.2218  0.8562
#>   0.8844 -7.4672
#>   2.4009 -11.1110
#>   0.5681  3.9024
#>   2.0086 -5.9228
#> 
#> (1,1,2,.,.) = 
#>   1.0283 -7.8774
#>   0.9488 -5.4806
#>  -0.3817  6.1782
#>  -7.9672  2.2322
#>   1.1391  5.3476
#> 
#> (2,1,2,.,.) = 
#>  -7.0302 -5.9904
#>  -15.1870 -2.4279
#>  -0.2613 -4.7066
#>   2.1727  0.6506
#>  -7.2376 -6.9688
#> 
#> (3,1,2,.,.) = 
#>   -0.6385   4.5094
#>   -2.1172   6.7111
#>    1.2294   3.0879
#>   -2.1425  -4.3473
#>   10.8541   6.7192
#> 
#> (1,2,2,.,.) = 
#>   -2.3247   4.8141
#>   -0.5857  17.4688
#>    1.5324  -7.6613
#>    6.1303  -1.3071
#>    0.6495   1.9523
#> 
#> (2,2,2,.,.) = 
#>   0.2665  4.2846
#>   6.2141 -5.3699
#>  -1.0734 -3.8871
#>  -2.3098  2.8314
#>  -2.0913 -10.5683
#> 
#> (3,2,2,.,.) = 
#>   2.0741  2.3961
#>   1.6505  1.3973
#>  -3.7137  0.3500
#>   3.3015 -6.8319
#>   3.9452  2.8915
#> 
#> (1,3,2,.,.) = 
#>   0.3765  7.4704
#>  -0.6238  1.2529
#>   0.6520  2.6082
#>  -0.4125 -1.8646
#>  -1.2206 -1.0096
#> 
#> (2,3,2,.,.) = 
#>  -16.7361   6.0327
#>  -14.1785  -2.3485
#>    4.1897   1.0602
#>   -3.7148  -0.8683
#>    5.5032   3.7983
#> 
#> (3,3,2,.,.) = 
#>  -2.2913 -1.0242
#>  -0.1949  3.0958
#>   8.6037 -6.3133
#>   7.6101 -4.9920
#>   4.7411 -11.8107
#> 
#> (1,1,3,.,.) = 
#>   5.5854 -2.7968
#>  -0.7211  3.7407
#>  -3.4941  0.1452
#>  -4.8603 -4.9176
#>  -5.6057  0.0670
#> 
#> (2,1,3,.,.) = 
#>    1.9008   7.0511
#>   -5.0117  -2.8580
#>   -2.6102   3.0189
#>    0.7731   0.1578
#>   -0.3252  13.2867
#> 
#> (3,1,3,.,.) = 
#>  -3.6498  0.2436
#>  -3.7511  3.7713
#>   3.0925 -6.1231
#>  -5.2023  5.1354
#>  -4.6123 -4.6121
#> 
#> (1,2,3,.,.) = 
#>   -0.8739   6.5956
#>   -3.3773  -5.2201
#>   -2.5145   1.7503
#>    1.9707  -8.6062
#>   10.7181   0.8793
#> 
#> (2,2,3,.,.) = 
#>  -4.0727 -2.5268
#>  -4.0274  0.4182
#>  -2.5292 -1.2610
#>   0.3872  1.8377
#>  -1.0432  4.7183
#> 
#> (3,2,3,.,.) = 
#>  -2.9777  6.5174
#>  -0.2439  3.2435
#>  -4.6745  0.9601
#>   4.8107 -1.5560
#>   3.5361 -3.8834
#> 
#> (1,3,3,.,.) = 
#>   1.2707  1.0076
#>  -9.4682 -5.6878
#>   3.3900 -4.4621
#>   5.0929  3.1561
#>  -0.9636  1.6640
#> 
#> (2,3,3,.,.) = 
#>  -9.0764  0.9988
#>  -8.8943  4.1230
#>   4.2488 -8.5396
#>  -6.0929 -2.5771
#>  -0.4581  4.7003
#> 
#> (3,3,3,.,.) = 
#>   1.9421 -5.9809
#>   3.0227  6.4231
#>   2.0785  1.2119
#>  -2.3858 -8.3997
#>  -7.3043 -16.2408
#> 
#> (1,1,4,.,.) = 
#>   5.5871  0.0724
#>  -0.9211 -1.5139
#>   2.8522 -5.5957
#>  -3.5227 -5.4352
#>  -9.9689 -5.9390
#> 
#> (2,1,4,.,.) = 
#>   6.5768  5.3043
#>  -3.8541  9.5912
#>  -1.5806 -0.0943
#>  -0.8956  1.6832
#>  -2.9333 -8.0291
#> 
#> (3,1,4,.,.) = 
#>   4.5893  9.0207
#>  -2.2989 -4.2598
#>  -5.7744 -8.3156
#>   6.8945  0.3199
#>  -3.5815  7.3833
#> 
#> (1,2,4,.,.) = 
#>  -5.4462 -1.2385
#>  -3.1519 -5.9413
#>   4.0939 -3.6897
#>  -2.2773 -3.6859
#>  -4.0197 -2.4303
#> 
#> (2,2,4,.,.) = 
#>  -11.7163   2.6541
#>    0.8376   2.1176
#>   -5.4378   4.2347
#>   -1.5888  -5.4482
#>   -5.9237  -2.0635
#> 
#> (3,2,4,.,.) = 
#>   4.8007 -10.5117
#>   1.4489 -2.0294
#>   3.9135  7.3048
#>   1.6503 -9.7240
#>  -7.8569  1.1108
#> 
#> (1,3,4,.,.) = 
#>  -7.4162 -1.1031
#>   5.1063  6.7199
#>   4.0016 -3.6399
#>   2.2488  2.5679
#>  -4.2789 -6.0855
#> 
#> (2,3,4,.,.) = 
#>   3.4099  5.6417
#>   1.2304 -8.2754
#>  -9.7752 -2.0893
#>  -3.4899 -0.9618
#>  -0.5475  1.3115
#> 
#> (3,3,4,.,.) = 
#>  -0.1344 -10.8859
#>  -1.8667 -0.1693
#>   4.2591  9.0988
#>   3.9502 -5.0591
#>   2.6254  1.7421
#> 
#> (1,1,5,.,.) = 
#>   -1.4406   2.2894
#>    3.9310   2.6438
#>    2.2587  -6.2077
#>    4.5898  12.6777
#>    6.2735   2.0686
#> 
#> (2,1,5,.,.) = 
#>  -1.2142  0.8348
#>   2.4895 -2.6613
#>   3.5994 -1.2268
#>   2.3945 -3.4609
#>  -4.0876  8.8493
#> 
#> (3,1,5,.,.) = 
#>   1.2492  4.8666
#>   1.4312  9.4923
#>   2.0753 -6.5545
#>  -6.1206  3.1949
#>  -5.6911  3.0633
#> 
#> (1,2,5,.,.) = 
#>   11.8472  -1.4907
#>    2.2849  -2.4078
#>    2.0723  -0.9115
#>    1.6150  -6.7548
#>   -1.3167   9.7604
#> 
#> (2,2,5,.,.) = 
#>   -0.7373  -0.1123
#>   -2.1284  -7.5241
#>   -0.8509   1.2806
#>    7.3625   5.8034
#>   10.8048   5.8343
#> 
#> (3,2,5,.,.) = 
#>   4.4169 -7.5077
#>   6.2526 -1.9890
#>   1.3288  1.9097
#>   1.2761  4.3698
#>   2.5885  5.6986
#> 
#> (1,3,5,.,.) = 
#>   0.1614  3.0715
#>  -3.6130 -0.8558
#>   1.7840  1.5174
#>  -2.8714  0.8049
#>  -4.4154 -6.5654
#> 
#> (2,3,5,.,.) = 
#>  -2.0639 -2.7726
#>   3.5464 -0.0129
#>   2.6387 -2.3682
#>  -4.1614  0.9227
#>   5.5742  0.7612
#> 
#> (3,3,5,.,.) = 
#>   2.0111  1.9548
#>  -5.7732  2.2125
#>   6.2816  2.6876
#>   5.8399  1.1591
#>   0.4935 -9.3194
#> [ CPUFloatType{3,3,5,5,2} ]