Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   0.5665  4.6707
#>   3.5746  8.1355
#>   5.4870  3.0339
#>  -2.4008 -7.5078
#>   5.8846 -4.1845
#> 
#> (2,1,1,.,.) = 
#>    2.0654   4.0419
#>   -4.3671  -4.3765
#>   10.1521  -0.8116
#>   -0.1816  -3.0199
#>    0.6656  -4.9893
#> 
#> (3,1,1,.,.) = 
#>   8.6207  2.4924
#>   1.4615 -6.3010
#>  -8.0386  4.8582
#>   0.7751  1.8816
#>  -1.8630  1.1176
#> 
#> (1,2,1,.,.) = 
#>  -6.0299  7.6651
#>   2.0031 -4.8063
#>  -12.1066 -0.1202
#>   6.1529 -7.2411
#>  -3.7772  5.5950
#> 
#> (2,2,1,.,.) = 
#>   -4.8881  -2.5800
#>    0.8094  -3.5587
#>   -2.6001  10.6678
#>   -3.1631  -8.9619
#>    2.7362  -3.5983
#> 
#> (3,2,1,.,.) = 
#>  -2.5083 -8.1692
#>   5.7977  3.0771
#>   3.4369 -10.7985
#>  -3.0127 -4.5961
#>   0.0704 -2.4183
#> 
#> (1,3,1,.,.) = 
#>  -2.5325  5.5493
#>  -3.7476  4.1371
#>   4.3417 -0.9237
#>  -4.4871  1.2462
#>  -1.0832  0.1587
#> 
#> (2,3,1,.,.) = 
#>   1.4903  2.0777
#>  -6.1162  5.9987
#>   3.5845  3.3041
#>   4.4716 -3.7569
#>   5.6711  5.7147
#> 
#> (3,3,1,.,.) = 
#>   2.1935  6.4387
#>  -3.4997  0.6281
#>  -7.9873 -4.4286
#>   5.3515 -4.2350
#>   1.0775 -3.3368
#> 
#> (1,1,2,.,.) = 
#>   5.6193 -4.0748
#>   1.9242  1.9409
#>   2.3719 -2.4550
#>   5.9963 -2.7474
#>   7.9202  2.2824
#> 
#> (2,1,2,.,.) = 
#>  -2.4796  7.9414
#>   0.6807 -1.7286
#>  -5.7195  1.2453
#>   2.1969 -1.6541
#>  -4.5841 -2.1633
#> 
#> (3,1,2,.,.) = 
#>   2.0019  1.2566
#>  -3.0804  3.2689
#>   1.5371 -8.0440
#>   9.9571 -3.3580
#>   7.3324  2.7023
#> 
#> (1,2,2,.,.) = 
#>  -3.4568 -3.1620
#>  -10.0626 -4.1922
#>  -5.3376 -4.3739
#>   1.6557  0.8278
#>  -2.1926  2.5537
#> 
#> (2,2,2,.,.) = 
#>   4.8847 -5.7936
#>   5.6233  3.7259
#>  -3.2647 -12.0651
#>  -4.1499 -1.3287
#>   1.8367 -3.4782
#> 
#> (3,2,2,.,.) = 
#>   8.6185 -6.1719
#>   2.8880  4.2468
#>   0.2826  6.6846
#>   0.5516  1.4976
#>  -3.0709  0.9730
#> 
#> (1,3,2,.,.) = 
#>   8.2100  5.5755
#>  -6.1693 -6.1230
#>  -0.6292 -5.7623
#>   7.5365 -1.2515
#>   7.6006 -4.8599
#> 
#> (2,3,2,.,.) = 
#>   9.9971  1.9912
#>   0.0104  3.3337
#>  -8.3728 -10.5856
#>  -1.3577  4.2733
#>   3.2276 -7.7098
#> 
#> (3,3,2,.,.) = 
#>   7.1332 -2.6983
#>  -4.1732  0.7347
#>   4.2084  1.6094
#>  -2.8934  0.2959
#>  -3.6532 -0.7526
#> 
#> (1,1,3,.,.) = 
#>  -3.9295  2.1628
#>   0.5688 -4.5369
#>   3.1830  1.9837
#>   6.4156 -1.5938
#>   1.0533 -0.5961
#> 
#> (2,1,3,.,.) = 
#>  -0.9433  1.8492
#>   6.7749 -3.9319
#>  -4.0906  7.6330
#>   4.3971 -5.3628
#>   5.6300  4.4504
#> 
#> (3,1,3,.,.) = 
#>   -1.1520   4.2689
#>   -4.5446  10.1079
#>   -9.4968  -0.4592
#>   -1.5261  -3.2082
#>   -0.1515   3.3532
#> 
#> (1,2,3,.,.) = 
#>    2.5446  10.8811
#>    2.6577   0.2402
#>   -3.0530  -2.0408
#>    5.0501  -6.5599
#>   -0.7297  -3.7726
#> 
#> (2,2,3,.,.) = 
#>  -15.1913   9.0637
#>    1.3451   3.0115
#>   -2.0624   1.1655
#>    4.2713   2.9985
#>    4.7365  -4.7846
#> 
#> (3,2,3,.,.) = 
#>   3.4787  5.9697
#>  -2.5484 -0.7507
#>  -8.1514  5.7021
#>   4.7598  6.6856
#>   3.1592 -5.0553
#> 
#> (1,3,3,.,.) = 
#>   7.3277 -6.4044
#>  -2.1692  6.3470
#>   2.0148 -2.7329
#>  -0.6055  3.7462
#>  -1.8985 -8.9990
#> 
#> (2,3,3,.,.) = 
#>  -0.4656 -3.9674
#>  -2.0111  1.4058
#>  -7.1171  0.4453
#>   6.8673 -4.1844
#>  -3.6216 -4.2420
#> 
#> (3,3,3,.,.) = 
#>  -4.1063 -7.2511
#>   1.4942  7.3610
#>  -9.6911  2.7466
#>  -3.2436 -3.2243
#>   4.7051 -0.2603
#> 
#> (1,1,4,.,.) = 
#>  -10.8671   3.9874
#>    5.2292   4.2299
#>   -3.7101  -8.8951
#>    2.1963   1.2888
#>    2.8746   0.7744
#> 
#> (2,1,4,.,.) = 
#>   7.2381  1.7528
#>   1.1890  0.8703
#>   8.0185  2.3313
#>   0.0514  2.5045
#>   2.2498  0.2309
#> 
#> (3,1,4,.,.) = 
#>   1.9136  5.1195
#>  -2.5939  1.1464
#>  -8.4254  3.7764
#>  -1.1451 -6.9906
#>  -1.7652  1.8945
#> 
#> (1,2,4,.,.) = 
#>   3.0137 -4.1380
#>  -9.7509 -0.8018
#>  -6.9653  1.9532
#>   0.8074  2.2981
#>  -5.3137 -6.7743
#> 
#> (2,2,4,.,.) = 
#>   2.6846 -5.3555
#>  -4.2456 -6.4054
#>   2.3545 -5.3559
#>  -3.0022 -4.4871
#>   5.5220  6.2112
#> 
#> (3,2,4,.,.) = 
#>   0.4799  1.7707
#>   4.9554 -5.9008
#>   5.3516 -4.4530
#>  -6.2509 -0.6878
#>  -6.1038 -7.5223
#> 
#> (1,3,4,.,.) = 
#>  -2.9632 -8.2271
#>  -6.3753  5.4741
#>   1.3251 -9.1539
#>   0.0935  4.9625
#>   4.5118  8.2055
#> 
#> (2,3,4,.,.) = 
#>  -7.3370 -0.2330
#>   1.5603  9.8136
#>   1.5331 -9.9215
#>   5.2926  1.2844
#>  -3.0401 -3.2529
#> 
#> (3,3,4,.,.) = 
#>   -0.4389   2.1170
#>   -3.1743   1.6413
#>    1.0762  11.2731
#>    9.9259  -6.0323
#>    2.8441  14.5390
#> 
#> (1,1,5,.,.) = 
#>  -6.7704  3.1231
#>   3.7632  3.9794
#>  -3.8934  3.5993
#>   1.5854 -4.9862
#>  -0.8416 -3.4794
#> 
#> (2,1,5,.,.) = 
#>   -1.7120  10.3147
#>   -4.1274  -2.4177
#>   -5.6083   3.3167
#>    1.2313  -0.3371
#>    4.0448   0.0549
#> 
#> (3,1,5,.,.) = 
#>    6.0636  10.0000
#>   -1.5494   4.3005
#>   -0.3156  -1.9874
#>    1.8566  -2.1273
#>   -5.3044  -1.1995
#> 
#> (1,2,5,.,.) = 
#>  -2.9894 -4.1329
#>   4.1016  1.8111
#>  -1.8422 -1.9255
#>   3.9524 -0.1068
#>  -3.2854 -3.7295
#> 
#> (2,2,5,.,.) = 
#>   3.3502 -1.6067
#>  -1.2560  6.6405
#>   0.8629  9.6598
#>   7.7196  2.2652
#>   0.7711  0.7030
#> 
#> (3,2,5,.,.) = 
#>   1.0895 -10.9194
#>  -6.0097 -4.8935
#>  -3.5493  1.3002
#>  -5.3833 -1.1504
#>   0.0893 -4.1874
#> 
#> (1,3,5,.,.) = 
#>   -0.5999  -9.8232
#>   -4.9075   0.8860
#>    4.8801  12.1747
#>    7.2655   5.8307
#>   -2.1658  -4.5190
#> 
#> (2,3,5,.,.) = 
#>   1.5824 -9.6312
#>  -1.7931 -3.6451
#>  -5.7958  0.7890
#>  -1.2505  3.1894
#>   0.9369  9.7620
#> 
#> (3,3,5,.,.) = 
#>    2.1966  -1.5974
#>   -3.0139   1.0524
#>    2.9046  -5.8181
#>   -4.1193  -4.5046
#>   15.5892 -14.7241
#> [ CPUFloatType{3,3,5,5,2} ]