Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   -0.4940  -9.2161
#>    1.6964   5.6273
#>    8.1563   3.8323
#>   10.1577   6.4075
#>   10.8842   0.5835
#> 
#> (2,1,1,.,.) = 
#>   -1.5216   8.1018
#>    0.8871  13.7185
#>    0.6848   1.9689
#>   -1.4194  -1.4012
#>   -3.6705   1.7456
#> 
#> (3,1,1,.,.) = 
#>   11.5719  -0.8694
#>   13.6515   4.4778
#>  -10.1591   3.8656
#>    1.0716   1.2461
#>    8.6176   0.5044
#> 
#> (1,2,1,.,.) = 
#>   4.5881  3.9021
#>  -3.1337  0.7641
#>  -6.7781  1.4605
#>   2.0881  6.4314
#>   1.5510 -5.3887
#> 
#> (2,2,1,.,.) = 
#>  -5.9582  1.0048
#>  -2.2981  3.7459
#>   1.9182  6.1685
#>  -8.6855  7.9156
#>   1.4061  0.5347
#> 
#> (3,2,1,.,.) = 
#>  -3.2292 -1.5918
#>  -3.2097  6.0182
#>   5.5262 -2.3041
#>  -8.0451  3.8534
#>   1.0027  0.6103
#> 
#> (1,3,1,.,.) = 
#>   -4.4169   2.3049
#>    5.5403  12.8065
#>   -4.2173   0.4357
#>   -1.4858  -0.0778
#>    5.6049  -0.2489
#> 
#> (2,3,1,.,.) = 
#>   0.2194 -0.9643
#>  -5.9894  3.6179
#>   4.2202 -3.4101
#>  -1.4789  3.2870
#>   5.4299 -3.1084
#> 
#> (3,3,1,.,.) = 
#>   7.4116  3.6093
#>  -2.0298  1.1393
#>   2.0428 -0.1127
#>  -2.7804  6.1492
#>  -0.9942  2.7427
#> 
#> (1,1,2,.,.) = 
#>  -3.6415 -2.3721
#>  -1.5322 -5.4560
#>  -1.0930 -0.1478
#>   6.6828  1.3282
#>   1.8203 -1.2577
#> 
#> (2,1,2,.,.) = 
#>   0.2765  0.5802
#>   2.0463  2.3181
#>   7.9410 -7.9059
#>  -2.1852 -4.2122
#>  -6.4102 -4.3553
#> 
#> (3,1,2,.,.) = 
#>   7.7379 -0.3637
#>   0.8687  1.3370
#>   4.4622  7.2542
#>   4.7449  2.6382
#>  -4.7433  3.2236
#> 
#> (1,2,2,.,.) = 
#>  -12.0410  -0.7815
#>   -0.3143  -3.7740
#>   -5.3588  -2.3949
#>   -0.1962  11.3009
#>    3.4620  -4.1507
#> 
#> (2,2,2,.,.) = 
#>  -10.4824   1.4575
#>   -5.7469  -6.1756
#>    5.5412   1.3685
#>   -4.4676   0.7655
#>   -8.7737   5.2888
#> 
#> (3,2,2,.,.) = 
#>  -8.4342 -3.8145
#>   1.1318 -7.7790
#>  -1.9571 -1.2592
#>   6.2227  7.4216
#>   0.5187  4.0033
#> 
#> (1,3,2,.,.) = 
#>   0.4994  3.3341
#>   1.0708 -1.2907
#>   1.6291 -12.1914
#>  -6.5956 -13.0465
#>  -4.1006  0.5697
#> 
#> (2,3,2,.,.) = 
#>   3.4166  3.0939
#>   4.6907 -6.0590
#>   6.1352 -7.9125
#>   2.0669  8.9465
#>  -4.4787 -1.3626
#> 
#> (3,3,2,.,.) = 
#>   0.1615 -6.0719
#>   2.3130 -2.3523
#>   3.8286  3.0982
#>   4.0661 -0.1951
#>  -0.8429  7.3103
#> 
#> (1,1,3,.,.) = 
#>   1.4621 -4.9929
#>  -5.6277  2.0744
#>   2.2013 -3.1144
#>  -4.9047 -4.3982
#>  -4.9158 -3.4195
#> 
#> (2,1,3,.,.) = 
#>  -4.2165 -7.4126
#>   1.4605  2.5431
#>   1.8166 -1.8854
#>   1.6960 -0.2001
#>   3.8423 -3.1945
#> 
#> (3,1,3,.,.) = 
#>    1.1422  -5.9551
#>   10.0559   8.3627
#>  -10.8562   2.5088
#>    4.2978   5.9558
#>    2.1281  -1.5993
#> 
#> (1,2,3,.,.) = 
#>   -3.0344  15.3783
#>    0.3399   2.0439
#>   -4.1796   4.2232
#>    5.6492  -4.4282
#>   -4.2593  -1.3188
#> 
#> (2,2,3,.,.) = 
#>   2.3720 -5.1832
#>   6.1296 -0.7297
#>  -0.4588 -3.8276
#>  -1.1014 -0.4439
#>   7.7604 -0.6264
#> 
#> (3,2,3,.,.) = 
#>   0.0478 -3.7499
#>   3.9332  0.9636
#>  -0.3666 -3.1233
#>   3.7668 -1.5615
#>  -6.9291 -1.7554
#> 
#> (1,3,3,.,.) = 
#>   2.9134  9.7364
#>   0.6708  4.4079
#>  -2.1989 -5.8824
#>  -0.0586  2.5904
#>   4.9945 -5.6211
#> 
#> (2,3,3,.,.) = 
#>   6.1073 -7.2683
#>  -0.4001 -4.9153
#>  -0.3558  3.7999
#>   0.8641  1.2973
#>   3.1867 -0.9580
#> 
#> (3,3,3,.,.) = 
#>  -4.9765  0.0733
#>   3.8983  4.5863
#>   0.0041  2.1443
#>   2.1421  2.8199
#>  -2.4740 -7.9438
#> 
#> (1,1,4,.,.) = 
#>   2.4906  3.0136
#>   0.8182  7.8543
#>  -8.5945  3.3681
#>   1.8979 -9.7592
#>   1.9501 -3.7281
#> 
#> (2,1,4,.,.) = 
#>  -2.3411  6.1591
#>   4.4010 -7.3815
#>  -9.1579 -1.2791
#>   0.8872 -1.9848
#>   4.6642 -0.4890
#> 
#> (3,1,4,.,.) = 
#>   5.2287  0.7678
#>   3.9830 -3.6319
#>  -1.8032 -3.1254
#>   1.7431  1.6613
#>   0.2589  4.0799
#> 
#> (1,2,4,.,.) = 
#>  -1.2660 -1.9864
#>  -4.6023  8.1299
#>   3.1738 -4.3730
#>   3.3230 -0.5513
#>  -3.2779  1.4687
#> 
#> (2,2,4,.,.) = 
#>  -8.9665  5.3559
#>   1.1850 -4.3773
#>   9.7727  2.1952
#>  -6.7316 -0.9335
#>  -1.1650 -8.1943
#> 
#> (3,2,4,.,.) = 
#>   0.4497  9.2806
#>   1.2121  3.7597
#>   0.8253  0.0568
#>   5.0589 -1.3721
#>   0.3898  0.8471
#> 
#> (1,3,4,.,.) = 
#>   -3.1620   3.2339
#>   -8.6805   4.0668
#>   -4.9654   8.9822
#>   -2.8723  -6.7886
#>   11.0570   3.7079
#> 
#> (2,3,4,.,.) = 
#>  -0.4565  5.4144
#>  -4.9227  0.7507
#>  -1.0770 -0.4754
#>  -9.3434  5.7442
#>   3.8448  4.5936
#> 
#> (3,3,4,.,.) = 
#>   0.8276 -4.1891
#>   0.2590  8.8877
#>   1.8251  3.0515
#>   3.6665  2.0245
#>  -4.6575  3.4559
#> 
#> (1,1,5,.,.) = 
#>   -2.4254  -2.1233
#>    7.4275  -0.2042
#>   -3.0799  -0.7439
#>   -7.7563  11.8731
#>    8.8087  -3.5168
#> 
#> (2,1,5,.,.) = 
#>   0.7126 -0.9245
#>   2.5958 -2.6771
#>  -1.5082 -11.4498
#>  -5.0399  2.2788
#>  -1.2218  7.0530
#> 
#> (3,1,5,.,.) = 
#>  -3.9625 -0.5612
#>  -5.0612  7.8809
#>   7.0091  1.2995
#>  -3.5154 -1.4217
#>   2.6349 -4.6922
#> 
#> (1,2,5,.,.) = 
#>    6.3978  -6.5413
#>    1.8317   6.7945
#>    3.8121  -1.4111
#>   11.1379  -5.5093
#>    4.0580   3.7424
#> 
#> (2,2,5,.,.) = 
#>    2.5517   3.5486
#>   -2.7667  -7.6918
#>   -2.4672   1.1463
#>   -2.2442  11.0000
#>   -0.9092  -3.4167
#> 
#> (3,2,5,.,.) = 
#>   -6.0773   3.0989
#>    4.4959  -2.0202
#>   -0.2786   3.9228
#>    9.2383  11.9286
#>    4.0910  -4.3850
#> 
#> (1,3,5,.,.) = 
#>   2.1852 -1.9210
#>   7.4026  1.7788
#>   0.3823 -1.8258
#>   0.5264 -1.3717
#>  -0.7918 -6.5605
#> 
#> (2,3,5,.,.) = 
#>   0.0151  6.4795
#>   1.8713  9.1754
#>   3.4914 -0.7373
#>  -4.4186  5.1439
#>  -2.1135 -7.5581
#> 
#> (3,3,5,.,.) = 
#>   1.1000 -11.4954
#>  -3.7423  1.3145
#>   9.1378 -1.1729
#>  -4.7550  7.8987
#>  -2.7389  4.8896
#> [ CPUFloatType{3,3,5,5,2} ]