Fft

Fft

torch_fft(self, signal_ndim, normalized = FALSE)

Arguments

self
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: FALSE

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to TRUE, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is torch_ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

if (torch_is_installed()) {

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)
# batched 1D FFT
torch_fft(x, 1)
# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

}

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -0.4542  2.7625
#>   3.5389 -8.3604
#>  -6.1493  2.3498
#>   0.5390 -3.2059
#>   0.1240  3.4884
#> 
#> (2,1,1,.,.) = 
#>   8.5467 -3.8740
#>   0.6077 -6.7279
#>   2.3105 -3.8631
#>   1.8239 -1.1472
#>   0.7949 -3.4914
#> 
#> (3,1,1,.,.) = 
#>   5.0885 -3.8094
#>  -1.9095  0.3443
#>  -9.7946  3.4984
#>   0.5096 -8.6538
#>   1.8563 -6.3566
#> 
#> (1,2,1,.,.) = 
#>  -2.0067  5.4405
#>  -4.4356  1.0321
#>   2.6340 -3.5594
#>  -0.6398 -5.4549
#>  -6.8691 -4.7308
#> 
#> (2,2,1,.,.) = 
#>  -10.7187   1.3853
#>    0.9343  -4.0556
#>   -9.7099   6.4042
#>    0.3021  -6.7301
#>    3.4487   2.8079
#> 
#> (3,2,1,.,.) = 
#>   3.0774  2.2036
#>   2.4140  4.9921
#>   4.1406 -3.8568
#>  -5.3231 -15.5160
#>   3.3530  2.5410
#> 
#> (1,3,1,.,.) = 
#>  -0.9754 -0.2065
#>  -2.5377 -2.7261
#>  -4.2094 -8.1411
#>   0.6086  2.1268
#>  -6.2971  5.7877
#> 
#> (2,3,1,.,.) = 
#>  -3.4029  1.6663
#>  -6.3625 -1.6552
#>   3.3041 -8.3189
#>   8.9531 -5.5423
#>  -1.1942 -2.9448
#> 
#> (3,3,1,.,.) = 
#>   7.0037  2.0335
#>   4.9229  4.3532
#>  -2.1362 -0.1693
#>   6.0184 -0.8980
#>   1.8110  3.3357
#> 
#> (1,1,2,.,.) = 
#>  -4.7952 -2.1140
#>  -3.8175  2.5847
#>  -6.0495  9.7531
#>  -2.2565 -0.3405
#>  -11.5547 -6.7817
#> 
#> (2,1,2,.,.) = 
#>   4.1394  2.8341
#>   5.2643 -2.6592
#>  -4.7602 -2.5962
#>   6.2405 -5.4166
#>  -9.5915  4.8499
#> 
#> (3,1,2,.,.) = 
#>   2.0557 -8.1355
#>   6.4111  8.1185
#>  -6.1405  6.6936
#>   3.1189  1.1755
#>   1.5754  2.6110
#> 
#> (1,2,2,.,.) = 
#>  -3.1409 -4.9458
#>   1.1644  0.2265
#>  -1.4126 -4.2623
#>   9.0104  0.6078
#>  -2.0481 -1.8685
#> 
#> (2,2,2,.,.) = 
#>  -1.4952 -1.2570
#>   5.5610  1.5609
#>  -6.4161  2.2876
#>  -1.8051 -0.3761
#>   2.6539  4.3449
#> 
#> (3,2,2,.,.) = 
#>  -0.7748  6.6679
#>   0.2316  7.5518
#>   5.1245  4.3019
#>   0.2382  0.1204
#>  -2.4745  1.2657
#> 
#> (1,3,2,.,.) = 
#>  -0.0022  2.9451
#>   4.0509 -4.8269
#>  -5.7987  6.6348
#>  -8.9609 -2.1075
#>   1.8041  4.7046
#> 
#> (2,3,2,.,.) = 
#>   -0.1038  -0.5838
#>   -1.5166   2.4285
#>  -10.2807  10.4550
#>    2.0513   5.7757
#>    3.9786  -2.8998
#> 
#> (3,3,2,.,.) = 
#>   -1.9071   6.5741
#>    2.2897  -1.5769
#>   -8.6233  19.4002
#>    0.7926   3.8249
#>   -4.3563  -1.4143
#> 
#> (1,1,3,.,.) = 
#>  -4.9856 -3.3698
#>   1.9406 -3.9610
#>  -2.2653 -0.5747
#>  -1.5889 -4.0037
#>  -5.1367  8.7530
#> 
#> (2,1,3,.,.) = 
#>   -6.3807  -0.4216
#>   -0.8966  -0.3943
#>    0.7035   4.8559
#>   -1.4947   3.7589
#>   -1.4374  11.9263
#> 
#> (3,1,3,.,.) = 
#>   1.1472  3.7160
#>   1.4928  7.4128
#>   5.4879 -2.8190
#>  -1.2266 -4.2955
#>   4.1317  1.5483
#> 
#> (1,2,3,.,.) = 
#>   5.4928  4.1962
#>  -5.5357  4.3974
#>   0.1962  2.1255
#>  -6.4028  8.8488
#>   5.6411 -11.4675
#> 
#> (2,2,3,.,.) = 
#>   4.4592  0.5581
#>   2.9216 -2.9231
#>  -10.4508 -5.8908
#>   0.5166  6.6552
#>   2.2413 -2.3384
#> 
#> (3,2,3,.,.) = 
#>  -2.3200  2.4347
#>  -7.9928 -0.7543
#>   1.4043 -1.7538
#>   4.3887 -0.7850
#>   1.5798  7.9320
#> 
#> (1,3,3,.,.) = 
#>   0.3598  4.8597
#>   5.3997 -0.8291
#>   1.6182  6.4560
#>  -4.0412 -2.0604
#>  -5.3286  3.2148
#> 
#> (2,3,3,.,.) = 
#>   7.6987  4.9007
#>   7.9020 -2.1932
#>  -3.7687 -5.1630
#>  -7.3750  0.8033
#>  -5.4754  7.9324
#> 
#> (3,3,3,.,.) = 
#>   0.9557  0.5381
#>  -5.3511 -7.2073
#>   6.2062 -2.4115
#>   9.6396 -3.4520
#>   6.6509 -7.2467
#> 
#> (1,1,4,.,.) = 
#>    1.3220  -7.5869
#>   -6.7002   4.5969
#>    2.9252   1.7297
#>    7.2020   0.6068
#>   13.8294  -1.6328
#> 
#> (2,1,4,.,.) = 
#>   0.5291  3.6611
#>   2.5358  2.7351
#>   1.9099 -1.1373
#>  -5.7598 -7.0873
#>  -5.3322 -6.4302
#> 
#> (3,1,4,.,.) = 
#>   0.7912  2.4924
#>  -2.5081 -7.8863
#>   2.6531  3.3621
#>   0.2168  2.5252
#>  -3.9718 -3.9398
#> 
#> (1,2,4,.,.) = 
#>   2.0678 -4.1828
#>  -2.3317  2.1605
#>   1.8425 -1.8036
#>  -2.3600 -1.7625
#>   2.5803 -0.4083
#> 
#> (2,2,4,.,.) = 
#>   4.1423  7.1647
#>  -1.2991  3.9615
#>  -2.8804  0.7320
#>   0.5405 -5.4450
#>  -6.5417  0.1442
#> 
#> (3,2,4,.,.) = 
#>    2.1904 -10.5987
#>   -6.0019   2.9451
#>   10.4787   8.5234
#>    4.2051   5.7612
#>  -13.0010   0.7807
#> 
#> (1,3,4,.,.) = 
#>  -5.9885 -7.9654
#>   5.6237 -1.7806
#>  -0.4926 -1.7048
#>  -0.5715 -1.6518
#>  -10.9420  2.6908
#> 
#> (2,3,4,.,.) = 
#>  -1.8901 -3.5469
#>  -2.3231 -0.0033
#>  -1.4247  1.2093
#>   1.4748  3.4100
#>  -7.9177  1.8113
#> 
#> (3,3,4,.,.) = 
#>  -3.5315  0.0271
#>  -0.6233  0.3096
#>   3.9069 -3.6904
#>   4.9048  1.5118
#>  -6.2757 -2.6916
#> 
#> (1,1,5,.,.) = 
#>   1.4799 -6.4956
#>  -1.2155  0.3255
#>  -7.7954 -0.9357
#>  -5.9716  4.4380
#>  -9.0794 -3.5732
#> 
#> (2,1,5,.,.) = 
#>   2.5038 -3.5253
#>   1.4063 -7.0058
#>  -0.4177 -2.4135
#>   2.7221  5.9849
#>  -3.4284  5.8031
#> 
#> (3,1,5,.,.) = 
#>  -0.5525 -7.0575
#>   1.7165 -1.5220
#>   0.6967 -2.2552
#>   8.6397  1.7221
#>  -9.1727  3.7125
#> 
#> (1,2,5,.,.) = 
#>    0.6627   2.7749
#>    1.2277   6.3292
#>  -10.4445   4.1807
#>    0.6249   3.4888
#>   12.5805   4.9652
#> 
#> (2,2,5,.,.) = 
#>   0.6798 -0.9390
#>   1.0238  8.6322
#>  -1.5867 -4.3453
#>  -5.8762  4.8749
#>   2.5303  2.9185
#> 
#> (3,2,5,.,.) = 
#>  -3.9090  0.5933
#>   4.8013  1.2237
#>  -5.1727 -1.7090
#>   1.0836 -5.8113
#>   0.3311 -11.8594
#> 
#> (1,3,5,.,.) = 
#>  -4.9257  3.0285
#>   2.4993  5.8339
#>  -2.2006 -1.4701
#>  -10.4100 -0.8071
#>  -5.1973  3.1384
#> 
#> (2,3,5,.,.) = 
#>   2.3806 -0.9498
#>   4.6529  2.6036
#>   4.2250  0.0661
#>  -3.7843 -12.8321
#>   2.8326 -1.2669
#> 
#> (3,3,5,.,.) = 
#>  -15.8541  -3.4800
#>   -1.6014   9.3139
#>    5.7523   4.1206
#>    3.0114   5.0920
#>   -2.3818  -7.7480
#> [ CPUFloatType{3,3,5,5,2} ]