Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.3523  2.6705 -0.2777  2.4044  0.3490
#>  -6.6455 -4.9170  3.5897 -2.7587  1.6472
#>   0.4345 -2.9674 -0.0112 -3.0171 -0.6049
#> 
#> (2,.,.) = 
#>  -1.4337 -1.7039 -0.4762 -2.4014 -1.7452
#>   0.5917 -0.7794 -2.3627 -2.1573 -3.4053
#>   0.2376  0.6762 -2.8528 -1.4173 -4.0121
#> 
#> (3,.,.) = 
#>  -3.9482  0.7397  1.3506  0.5112  0.4305
#>   1.1008  1.8285 -0.0168  0.8150  0.5168
#>   1.7841 -0.5175  1.2347  2.2275  2.9030
#> 
#> (4,.,.) = 
#>   0.7495 -2.5154 -0.4975 -3.4962 -1.2801
#>  -1.4299  0.4525  0.2726  0.3180 -0.0902
#>  -1.0822 -0.7741  3.9784  1.6111  5.2096
#> 
#> (5,.,.) = 
#>  -0.0783 -1.8290  0.6622 -0.2340  0.8436
#>   0.2054 -1.8169  2.1320  0.5583  3.0545
#>  -1.6283  3.7571  0.1242  1.5152 -0.1675
#> 
#> (6,.,.) = 
#>   2.2825 -0.5559  0.1114  0.9450  1.2666
#>   1.8791  2.2993 -2.3116 -0.8777 -2.6631
#>   0.8125  1.1760 -0.3168  2.3302  0.5202
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]