Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -1.2643 -0.7788 -2.7737  5.7298  2.3719
#>   2.4890  0.6918  0.3022  2.3339 -1.7301
#>   1.7037  0.7974  0.6570 -0.0111 -0.3945
#> 
#> (2,.,.) = 
#>  -4.1850 -0.6582 -2.6826  1.3492  1.4278
#>  -0.2409  0.8300  1.2109 -3.9962 -1.9565
#>   3.0714  0.7320  0.8258  1.5358 -0.1297
#> 
#> (3,.,.) = 
#>  -0.8558 -0.3055 -0.7542  0.9995  0.1580
#>  -0.9736  2.2068 -0.0192 -3.6943  0.4909
#>  -1.9552 -0.3294 -1.5302  1.4647  0.0024
#> 
#> (4,.,.) = 
#>  -1.6764  0.6105 -0.2673 -2.3274 -0.0756
#>   0.9726 -0.4710  1.3594 -1.6635  0.3816
#>   0.0680  0.4090 -1.0876  2.2780  1.1629
#> 
#> (5,.,.) = 
#>  -0.5785 -0.0512  1.0340 -3.2124 -1.3339
#>  -1.4959 -0.6073  0.5417 -2.8129 -0.5300
#>  -0.1862  0.2941 -0.5601  0.9484 -0.5442
#> 
#> (6,.,.) = 
#>   3.1933  1.9776  3.2224 -5.2617 -3.0172
#>   3.1556  3.6237 -1.4071  4.5237 -1.0186
#>  -0.3620  0.6285  1.8153 -5.3518 -2.9632
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]