Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -2.37961
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.3198
#>  0.0653
#>  0.3550
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.2162  0.5417  1.3272
#>  0.7006 -0.2425 -0.5578
#> -1.0573  0.4057 -0.6394
#>  1.6030  1.8447 -1.8793
#>  0.0370 -0.0203 -1.7582
#> -0.3430 -2.1128 -0.2493
#> -0.3039 -3.4963 -0.0323
#>  1.8722  0.8350 -1.8420
#>  0.4691  0.9965 -0.7580
#>  1.4223  3.4256 -3.3550
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   1.1113 -1.1552  0.8760 -0.8790 -2.8580
#>  -0.2041 -0.6704  0.3028  1.1685 -1.3819
#>   3.9735 -1.1724 -2.3308  1.2773  2.3590
#> 
#> (2,.,.) = 
#>  -1.8590 -1.1571  1.6596  3.4932 -0.2053
#>   0.7907  1.5118  2.3988  1.1325 -1.4084
#>   0.7586  0.7758 -0.3884 -4.7001 -1.7690
#> 
#> (3,.,.) = 
#>   1.3623 -2.2671  0.6622 -4.7628  3.8033
#>   1.6976  0.1071  2.5860 -2.4877  3.6707
#>  -0.6107 -2.2086  0.8495  1.6411 -0.0588
#> 
#> (4,.,.) = 
#>  -0.6850 -2.4919  1.6825  0.0037 -0.2940
#>  -0.3664  0.6351 -0.2900  1.3225  1.2269
#>  -2.3694 -4.7235  3.2173  1.7753  1.4479
#> 
#> (5,.,.) = 
#>  -1.1360  0.8370 -0.0065 -1.3364 -6.1653
#>  -0.4376 -1.4373  1.2980 -1.2722 -3.8815
#>  -2.4885  2.9614 -3.5468 -1.5805 -1.4790
#> 
#> (6,.,.) = 
#>   0.0274 -1.8097  0.1383  0.7062 -0.2213
#>  -0.7995 -2.1534  1.1278  0.4569  1.6627
#>  -0.4890 -1.9242  3.1078  1.0225 -0.2666
#> 
#> (7,.,.) = 
#>  -0.8293 -0.7764  0.7245  0.9477 -0.4161
#>  -0.7256 -0.7374  0.0962  0.3372  0.3064
#>   1.4338 -1.3974  1.2873  0.5058 -0.3121
#> 
#> (8,.,.) = 
#>  -0.9770  1.1870 -6.5508 -1.2618 -1.3953
#>  -1.5671  1.2927 -3.4435 -0.5430 -1.0933
#>  -0.3111 -0.6995 -1.8576  0.4555 -0.1343
#> 
#> (9,.,.) = 
#>   1.1739 -0.6017  0.5432 -0.7958 -2.5355
#>   2.8817  0.4214 -0.1819 -1.2467 -2.2137
#>  -0.8617 -0.8101  0.6802  0.0556 -2.7424
#> 
#> (10,.,.) = 
#>  -1.6688 -2.1020  1.6298 -1.4685 -0.7954
#>  -0.0248  0.4468  0.0037 -0.2339  0.3715
#>   0.0273  0.6181  0.3543 -1.4158  0.0447
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.2903 -0.1713  1.5687  0.2440  0.8829
#>  -1.1501  0.0086 -0.6468 -0.3439 -0.7836
#>  -0.4925  0.2729 -2.5634 -0.7050 -2.1906
#> 
#> (2,.,.) = 
#>   0.3274 -1.6134 -2.5264 -0.1420 -0.8973
#>   0.5083 -3.8758 -1.4592  0.2918  0.0364
#>   0.7792 -1.2554 -2.9096 -0.2987 -1.5335
#> 
#> (3,.,.) = 
#>   0.0405 -2.4622  0.6054  0.3894  0.7997
#>   0.1680  1.7805  1.3400  0.1231  0.7338
#>  -0.2414 -1.4517  0.0232  0.1096  0.1513
#> 
#> (4,.,.) = 
#>  -0.1626  1.7795  4.2497  0.5559  2.2954
#>   0.3687  0.7221 -0.0990  0.0198  0.1044
#>   0.4858  1.7425  0.5109 -0.0192  0.1279
#> 
#> (5,.,.) = 
#>  -1.9948  3.0173  0.5147 -0.6715 -1.0102
#>   0.5201 -2.5219 -2.8538 -0.1559 -1.2146
#>  -0.3154 -2.8353  0.9963  0.5307  1.3404
#> 
#> (6,.,.) = 
#>  -1.2946  3.9233 -0.2768 -0.7591 -1.3455
#>   0.1517 -1.3708  4.3827  0.9880  3.0092
#>   0.4771 -2.0279 -1.7461  0.0110 -0.5293
#> 
#> (7,.,.) = 
#>  -0.9860  2.3166  0.7560 -0.3039 -0.2300
#>   0.4438  0.0314  1.0753  0.2857  0.8275
#>  -0.6210  0.5084  3.2703  0.4971  2.0039
#> 
#> (8,.,.) = 
#>   0.4506  1.0349  3.2408  0.6333  2.2187
#>   2.0914 -2.5862  1.6400  1.0634  2.3967
#>   0.3238  0.9694 -0.4193 -0.0741 -0.1474
#> 
#> (9,.,.) = 
#>  -0.3307  2.5837 -2.5615 -0.9645 -2.6573
#>   0.3015  3.3034 -2.1871 -0.7720 -2.0481
#>   0.5390 -1.2913 -0.7913  0.0995 -0.1046
#> 
#> (10,.,.) = 
#>   0.2949 -0.5820 -1.6155 -0.1612 -0.7567
#>  -0.3058 -1.8320 -0.3674  0.0708 -0.0356
#>  -0.5654 -1.5082 -1.8227 -0.2263 -0.8751
#> [ CPUFloatType{10,3,5} ]