Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.1444  1.8934 -0.2313 -1.3341  0.1848
#>  -1.4181 -2.0209 -2.1780 -5.7290 -1.7862
#>   1.1241  7.1974  1.8319  2.1853  1.1714
#> 
#> (2,.,.) = 
#>  -2.4868  1.6245  0.6686  0.9076  0.9681
#>   0.7048 -1.0987  0.4870  3.4667  1.6040
#>   0.9822 -0.5871  0.3562  1.7710  0.4777
#> 
#> (3,.,.) = 
#>  -1.6559 -0.7625  1.1425 -0.1509 -1.6771
#>  -2.0375  2.3577  2.2445  4.8269  1.9810
#>   2.6247  0.6088 -0.1339  1.2305  0.7297
#> 
#> (4,.,.) = 
#>   2.5764 -1.3335  0.3058 -0.0145 -1.6303
#>  -0.1065 -1.4027  1.6155  2.7117 -0.6332
#>   2.8413 -0.9355 -2.0780 -3.1378 -0.5540
#> 
#> (5,.,.) = 
#>  -0.5339  1.9766  0.2453 -0.2907  0.1961
#>  -0.8401  2.3325 -1.4782 -1.0996  2.3492
#>   0.2835 -1.0054  0.1086  1.8781  0.9936
#> 
#> (6,.,.) = 
#>  -0.7578  3.8154  2.5283  5.2691  2.0649
#>   0.6690  3.0374  1.0334  2.4780  1.3505
#>  -1.5939  1.2891  0.3068 -0.4993 -0.0175
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]