Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> 3.73174
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.4115
#>  3.5397
#>  0.0928
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  2.5827  2.6949 -1.3013
#>  0.7192 -1.5363 -0.7182
#> -0.3708 -0.0438  0.8197
#> -0.7875  0.8994 -1.8001
#>  1.3229 -1.3084  0.3698
#> -1.9201  2.1622 -0.1773
#>  0.5786 -1.5993 -0.8559
#>  3.1901  0.9263 -0.7546
#> -0.9247 -2.9735  1.7972
#>  0.6887  0.4814  0.2335
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -1.0245  0.8449 -0.7120 -0.1708  0.0676
#>   1.4217  3.3911  0.6550  1.7426 -2.1456
#>   0.4247 -1.3422  0.0493 -0.5735  0.0949
#> 
#> (2,.,.) = 
#>  -0.2540 -0.9837  0.8524 -1.0587  0.3196
#>  -1.1599 -1.2633 -0.7615  0.6941 -1.5605
#>   3.2021  0.7883  1.1156 -1.6008  1.0204
#> 
#> (3,.,.) = 
#>  -2.4101  1.5616  1.3150 -3.4160 -1.7665
#>  -0.0040  0.3949  2.7630 -3.1926 -1.0730
#>   0.8214 -0.1401  0.6880 -2.8071 -2.3101
#> 
#> (4,.,.) = 
#>  -0.3839  1.5893 -3.6436 -4.7059  2.9374
#>  -4.9267 -0.3690  4.8539 -0.6227 -4.8704
#>   0.1376  0.8120 -3.7900 -5.0406  3.0316
#> 
#> (5,.,.) = 
#>   0.6562  5.1001 -2.3387  0.0102 -0.8945
#>   0.6922 -3.7446 -0.6112  0.6796  2.0216
#>  -0.0199  5.6328  0.8415  0.3285 -1.3051
#> 
#> (6,.,.) = 
#>   0.9286  0.2641  1.0525  0.9980  0.4851
#>  -0.5903 -2.1345 -0.3125  0.2031  1.0148
#>   2.9825  4.4798 -2.6958  1.1492  0.6500
#> 
#> (7,.,.) = 
#>  -0.2238 -0.5147  0.4704  0.3624  0.2358
#>  -3.3406 -0.2978  0.4195  3.5891 -2.5762
#>  -0.9460 -2.3469 -1.6619 -3.2283 -0.6149
#> 
#> (8,.,.) = 
#>   0.0699  1.0762  4.0092  0.1633  1.8639
#>   2.7599  3.4194  0.5122  2.0457 -0.5986
#>   2.5415  3.0959 -2.1131  2.0791 -1.8277
#> 
#> (9,.,.) = 
#>  -0.0951 -0.1282 -0.7652  1.8993  0.5640
#>  -0.6378 -0.8623  0.1689 -0.9280  1.2023
#>  -2.1805 -1.3852 -2.3629 -1.6894  3.2962
#> 
#> (10,.,.) = 
#>  -0.5315  1.7623 -0.3792 -0.4425  0.8818
#>  -0.8631  0.2120 -0.5367 -0.8567  1.3216
#>  -0.9344  0.9088  0.5899 -1.1179 -1.7244
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -4.4810  3.2554  6.0841 -1.0117  3.9206
#>   0.8889 -0.2357  0.8999  2.2952 -1.8829
#>  -2.6604  2.5916  6.1432  0.9864  0.7025
#> 
#> (2,.,.) = 
#>  -2.3561  1.3166  3.0731  1.1236  3.2472
#>  -1.3011  1.9987  5.0089 -0.1287 -0.2196
#>   2.7747 -1.0899 -1.0301  1.5347 -5.1708
#> 
#> (3,.,.) = 
#>  -0.3180  0.3614  0.6155 -1.6093  1.8880
#>  -1.5611  0.7715  1.4717  1.0149  1.0426
#>  -4.0554  2.2548  3.0323 -1.8050  4.5958
#> 
#> (4,.,.) = 
#>  -1.2326  0.3929 -0.2714  1.1295 -1.8483
#>  -0.7872 -0.2010 -0.8013  0.1374  2.5753
#>   0.9791 -1.1289 -1.8809  0.9289  0.7395
#> 
#> (5,.,.) = 
#>  -1.0499  0.5348  0.3582 -0.2332  0.0163
#>  -2.6341  1.5109  2.6744  1.0204  1.2279
#>   3.6258 -2.2115 -4.6907 -1.4654 -3.3297
#> 
#> (6,.,.) = 
#>  -2.8084  0.3528 -0.8032 -0.6248  5.9909
#>   3.3837 -1.6526 -2.4162  0.2684 -3.9548
#>   1.5162 -1.0458 -2.6265  0.2145 -3.4753
#> 
#> (7,.,.) = 
#>   1.6609 -1.1019 -2.5406 -0.0657 -2.7526
#>   2.6352 -1.0475 -0.9002  3.0114 -6.9826
#>  -2.0996  2.1623  5.6533  1.7140  0.3396
#> 
#> (8,.,.) = 
#>  -5.1709  3.1816  6.9473  1.6506  5.9466
#>  -3.4099  2.6091  6.6739  2.5913  2.6344
#>   1.3836 -1.2238 -3.3656 -1.3986 -0.8137
#> 
#> (9,.,.) = 
#>  -0.6961  0.2595  0.7204  1.7522 -0.4876
#>   1.1652 -0.8838 -2.5106 -1.8549 -0.1696
#>   0.6479 -0.5438 -1.8835 -0.3433 -2.1348
#> 
#> (10,.,.) = 
#>  -1.5713  0.8517  2.5633  1.2122  3.3083
#>   4.3215 -3.4144 -7.0254  1.2452 -4.8852
#>   4.0163 -2.3574 -4.9049 -1.0801 -4.6177
#> [ CPUFloatType{10,3,5} ]