Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# \dontrun{

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -2.01064
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.8498
#> -0.1725
#> -0.2413
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.3051  0.8335  0.7817
#> -1.7290 -0.1315 -1.0013
#> -2.2799 -2.0186 -1.7212
#> -2.0462  2.6530  2.2152
#>  1.0726 -1.8564 -0.0805
#>  2.5083  0.1464 -1.9015
#>  0.1432  1.3745  0.8548
#> -0.5524  0.7222  0.1316
#> -1.5975  0.3861 -2.5685
#> -3.1016  2.3700  0.1975
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   2.7383  3.7776  1.1984 -0.0462  5.9024
#>   0.1247 -2.9988  0.8401 -1.2394 -2.9948
#>   3.0933  3.5992  3.5134 -3.3567  6.0899
#> 
#> (2,.,.) = 
#>  -0.9786 -0.3967 -0.3440  0.3824 -1.2635
#>   0.4944  0.7149  0.5734 -1.1592  0.6260
#>   1.3268  0.8352  1.9060 -3.7034  0.1600
#> 
#> (3,.,.) = 
#>   0.4321  0.9564  1.6369 -0.1843 -0.3877
#>  -1.4475 -2.3336 -2.1501 -1.6377  3.0016
#>  -1.9856  0.2134  1.8436  1.6812  1.5892
#> 
#> (4,.,.) = 
#>   0.7306 -0.1213  0.6505 -2.7818 -0.1171
#>  -1.4915  3.2909  0.5003  2.5146  1.3735
#>   0.0837 -0.1117  0.8951  3.6226  1.1665
#> 
#> (5,.,.) = 
#>  -1.9671  0.1609  2.2518  0.7587  2.6455
#>  -0.5965 -1.3687 -2.4585 -1.6623 -3.6542
#>   0.0899  1.6690  0.2659 -2.6994  2.8543
#> 
#> (6,.,.) = 
#>  -4.5552  0.7339 -2.3152 -5.9254 -1.6420
#>  -0.8628  0.0829 -0.6564  0.9928 -0.0438
#>  -4.1770  0.7927 -2.3129 -4.7196 -0.5457
#> 
#> (7,.,.) = 
#>   1.3550 -0.9541 -1.7768  2.5931  1.6567
#>   4.4144  1.4763  1.1692 -2.8514 -1.1695
#>   4.2236  0.0365 -0.3523  1.0364  0.0855
#> 
#> (8,.,.) = 
#>  -3.3315  1.4609 -1.3792 -2.9510 -0.7443
#>   1.3569 -2.2547  0.7054 -2.1291  5.2801
#>  -0.0102  1.3483 -0.3866  2.5474 -3.7250
#> 
#> (9,.,.) = 
#>   0.8332  2.8945 -1.7215  1.0412  1.7159
#>  -1.0112 -5.9806  1.3106  3.5058 -2.5003
#>   0.3979  0.5793 -0.9296 -2.5669 -3.1548
#> 
#> (10,.,.) = 
#>   1.3155 -0.3605  0.2988 -0.0092 -0.7717
#>   0.6723  1.2307 -1.8691  0.1663 -1.7657
#>  -1.2338  0.4575  0.0627  1.7217  2.0104
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -5.1244  2.2971 -1.6275  2.7803 -4.4830
#>   1.8873 -3.6560  1.6412  0.0507  0.1216
#>  -0.5433  1.9666 -1.4803 -0.5111  0.0638
#> 
#> (2,.,.) = 
#>   0.0601 -4.0015  2.7077  1.7992 -1.7779
#>  -1.6433  1.4184 -1.3365  0.5739 -1.9782
#>  -3.4216  6.9532 -3.3050 -0.2836  0.0183
#> 
#> (3,.,.) = 
#>  -1.8677 -0.3711  0.3040  1.6625 -2.7921
#>   3.2992 -0.5994 -1.6079 -2.2177 -1.6740
#>  -3.7891  4.2637  2.9312  2.2233  0.8338
#> 
#> (4,.,.) = 
#>  -2.1277  3.2238  0.0312  0.5080  1.3242
#>   2.2424 -0.8814  0.2856 -1.2700  1.1283
#>  -1.0811  2.6808 -0.7240 -0.2231  1.0720
#> 
#> (5,.,.) = 
#>  -3.4333  2.0806 -0.9037  1.4697 -0.1231
#>   2.3334 -3.4847  1.1505 -0.3664 -0.2367
#>   1.2297 -2.4053  2.6418  0.6376 -1.1939
#> 
#> (6,.,.) = 
#>  -2.2066  0.7770  1.6664  1.7391 -0.3200
#>  -2.7109 -1.6160 -1.9334  1.4483  0.3787
#>  -1.2449 -0.9002 -1.0765  0.8913 -1.9713
#> 
#> (7,.,.) = 
#>   0.2789 -1.0134 -0.5889  0.0808 -1.8249
#>   0.6334 -0.8853  2.2413  0.6094 -1.5061
#>  -0.1358  0.5410  0.6620  0.1202  0.3539
#> 
#> (8,.,.) = 
#>   0.5331  0.8902  1.7452 -0.4616  3.8193
#>   0.4330  2.0916 -0.6593 -0.9295  0.1894
#>   0.8105 -0.3216  0.7608 -0.4534  2.0104
#> 
#> (9,.,.) = 
#>   2.2947  1.8882  2.8289 -1.2676  2.0636
#>  -0.6974 -2.0088 -1.8360  0.1831  1.3406
#>   4.7011 -3.7311  0.0230 -1.9564 -1.7259
#> 
#> (10,.,.) = 
#>   2.7739 -3.8386 -1.8082 -1.4422 -0.0495
#>   1.3869 -3.8928  0.8670  0.1969  0.0855
#>  -3.0783  1.1996 -2.8824  0.9550 -0.9141
#> [ CPUFloatType{10,3,5} ]

# }