Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.0241494
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.2363
#>  3.0268
#>  0.1035
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -2.7407  2.5049  0.0359
#> -0.2428  1.9005 -0.0311
#>  1.5064 -1.1403  4.7911
#> -1.1711  1.5313  4.8539
#>  1.0677 -1.4385  3.5871
#> -3.6176  2.1055 -0.4337
#>  2.7659  3.3263 -1.1249
#>  3.1125 -1.0300  0.8533
#> -7.2295 -3.2742  1.9161
#>  1.7768 -0.6715  2.4132
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   2.1537 -0.9727  0.0868  3.0558 -0.9135
#>  -2.6331 -8.2079  4.6051 -6.7919  5.9808
#>   0.4733 -2.2162  1.0598  2.4622 -0.4438
#> 
#> (2,.,.) = 
#>  -3.2293 -1.7143 -2.6607 -0.0832 -1.6717
#>  -2.5618  1.4231 -0.5654  0.1853 -0.0330
#>  -1.1593 -0.6543  0.1057  0.7253  0.2542
#> 
#> (3,.,.) = 
#>   1.8733 -2.5163  3.7143  1.6841  0.1385
#>   1.2127 -2.9773  2.3041  2.5720  1.3138
#>  -0.0171 -0.9660  3.6009  0.7866 -1.5337
#> 
#> (4,.,.) = 
#>   0.3547 -3.9708 -2.5380 -1.0752 -5.3527
#>   1.7091 -2.9129  0.9074  1.0743  1.0275
#>   1.1593 -1.8107  0.8382  0.4825  1.2793
#> 
#> (5,.,.) = 
#>  -0.5769 -4.2236 -1.0838  0.4376  2.2441
#>   0.7824  1.9849 -1.4272  0.1560 -0.0576
#>   0.7344 -2.5135 -0.5494 -0.3676  0.4109
#> 
#> (6,.,.) = 
#>  -1.7212  2.1676 -0.9803  0.0793  2.8358
#>   0.9456  0.0361  0.5637 -1.1537 -0.1105
#>  -0.5617  1.4245 -0.3344  0.0940  1.5494
#> 
#> (7,.,.) = 
#>   2.5519 -0.1848  2.4864  1.5132  1.0586
#>   0.6640 -0.3814 -0.6928  1.5510  1.9655
#>  -0.8183 -0.3852 -2.2143  1.1407  2.0630
#> 
#> (8,.,.) = 
#>  -0.0168  1.3795 -1.1796  3.2791  0.9693
#>  -0.2317  0.4384 -1.5643  2.2430 -0.6618
#>  -0.0996  0.6698  0.4613  0.2960  1.1599
#> 
#> (9,.,.) = 
#>   7.9252  0.2456 -7.1531 -2.9645  3.2929
#>   2.1319  0.3302 -2.2897 -0.6544  1.2224
#>   4.0157 -1.7977 -1.3135  2.6495  0.4928
#> 
#> (10,.,.) = 
#>   0.1402  0.4785 -3.5549 -1.0089  0.2646
#>   1.8108  0.0249 -0.4183 -0.9522  1.3001
#>  -0.6563 -0.2119 -0.7839 -0.4225  0.3819
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   3.3780 -1.0720 -1.3765 -5.1701 -0.5722
#>  -1.2270 -0.7055 -0.0240 -1.6081  0.2402
#>   3.7207 -0.8630 -1.5693 -3.1595 -0.2475
#> 
#> (2,.,.) = 
#>  -3.5109 -2.6442 -0.8076  1.2871  1.0722
#>   0.7154 -0.0828 -0.1345  2.5352 -0.6859
#>   2.2720  2.0566 -0.0048  1.3469  1.0371
#> 
#> (3,.,.) = 
#>   1.7905 -3.0443 -0.9543 -4.1480 -3.4618
#>  -5.6131  2.7879  2.6081  1.3717  2.4207
#>  -0.2866  0.0182  0.0553 -1.1232  0.2584
#> 
#> (4,.,.) = 
#>   2.8548 -0.5644 -0.4335 -1.3244 -2.2084
#>   0.7430  0.1274  0.0114  3.4985 -0.8686
#>   1.3035  1.0213 -0.9364  4.6209  2.4461
#> 
#> (5,.,.) = 
#>  -4.1717  1.7929  2.3653 -0.7239  0.3967
#>   0.2925  2.7194  1.1932 -0.4505  0.8761
#>   2.4023 -0.9088 -0.0130 -5.5450 -3.0284
#> 
#> (6,.,.) = 
#>   6.0704 -1.7497 -2.1878 -3.5899 -2.0988
#>   0.4229  0.6318  0.6280 -3.0756 -0.7544
#>   0.6998 -0.0218 -1.0562 -0.8221  2.2329
#> 
#> (7,.,.) = 
#>  -5.3377 -0.5594  1.1731  3.1231  0.5482
#>  -5.5583 -0.7444  1.3053  0.8009  0.3432
#>   0.4578 -4.8776 -2.4537 -2.8592 -1.2562
#> 
#> (8,.,.) = 
#>   4.1146 -2.6102 -2.8294  1.8040 -0.5032
#>  -4.8004 -0.3467  0.8817  3.2605  1.1562
#>   1.9305 -1.0591 -1.5835 -2.6062  1.0635
#> 
#> (9,.,.) = 
#>  -0.9492 -0.1283  0.9508 -1.7143 -1.6992
#>   6.1829  3.3914 -0.0867  2.4016 -0.3277
#>  -2.9519 -2.7414 -0.7203  3.1738 -0.0198
#> 
#> (10,.,.) = 
#>  -3.2094  2.7451  2.9043  0.2943 -0.5282
#>   1.9241  0.7552 -0.0242  4.1561 -0.9324
#>   2.2007 -0.6206 -0.1254 -4.8430 -2.1609
#> [ CPUFloatType{10,3,5} ]