Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.6556 -1.4940 -1.1538  0.6211  2.9385
#>   1.1426  0.6408 -0.9266  0.5678 -1.9136
#>  -1.1025  1.1658  1.6736 -0.8056 -1.9601
#> 
#> (2,.,.) = 
#>   0.5522 -0.5421  0.6067  0.3333 -0.7652
#>  -2.3665  3.1543  1.0085 -1.6725 -2.0849
#>   0.8027 -4.0446 -0.6147  0.9108  5.5314
#> 
#> (3,.,.) = 
#>   0.5812  0.7414 -0.0904  0.2657 -1.7143
#>   1.8026  1.7792 -1.4488  0.8381 -3.9641
#>   0.7097  0.2003 -0.3075  0.3401 -1.3855
#> 
#> (4,.,.) = 
#>   0.1020  2.0735  0.6998 -0.1427 -3.5905
#>   0.1597  0.2086  1.0326  0.0271 -1.6888
#>  -1.6120  1.8429 -0.2242 -1.1424 -0.5404
#> 
#> (5,.,.) = 
#>   0.0990 -1.8232 -0.0468  0.2083  2.2429
#>   0.3417 -0.0848  0.2677  0.2335 -0.3149
#>   1.8157  0.3540 -1.3463  0.9952 -1.9785
#> 
#> (6,.,.) = 
#>  -2.1095 -2.4062  1.1770 -0.8772  6.0952
#>   0.2243  1.6711  0.1035 -0.1176 -3.4111
#>  -1.4390  2.6204  1.5148 -1.1242 -3.2844
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]