Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   0.6752  0.6343  0.1261  0.6579  0.2426
#>  -0.0305  1.0896  1.0189 -0.9649  0.3235
#>  -2.0484  1.5087  2.8357 -1.2280  2.6115
#> 
#> (2,.,.) = 
#>   0.8323 -0.7302 -1.5227 -0.1321 -1.6173
#>  -0.0778  0.4187  0.4003  0.3667  0.6518
#>   1.0180  2.8191  2.4117 -1.2179  0.6263
#> 
#> (3,.,.) = 
#>   4.1507  0.8266 -3.1833  2.6300 -2.2084
#>   2.7421 -0.5495 -2.8164  2.5538 -1.7845
#>   0.7421  0.7768 -0.7787  1.8799  0.8418
#> 
#> (4,.,.) = 
#>  -1.8953 -1.4606 -0.7091 -0.2861  0.4259
#>   4.5689  1.5038 -1.6069  1.4254 -2.5641
#>   1.6301  0.3128 -1.6116  4.1578  1.0227
#> 
#> (5,.,.) = 
#>   1.6633 -0.0319 -1.8567  1.0804 -1.2252
#>   0.3281 -3.2238 -2.0250 -0.6907 -3.1894
#>  -1.8451 -2.1874 -1.2366  1.0216  0.6387
#> 
#> (6,.,.) = 
#>   1.6954 -1.1157 -2.0229  0.4779 -2.3466
#>   0.7501  0.3583 -0.5328  0.0332 -0.5396
#>   0.3005  2.0894  2.7034 -4.6540 -1.3034
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]