Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   1.4141  1.4001 -1.1872  4.4812 -4.5123
#>  -0.6182 -0.5955  0.2930  0.0105 -0.2454
#>  -1.1772 -1.2618 -2.2089 -0.6403 -3.9477
#> 
#> (2,.,.) = 
#>  -0.3450  0.8121  0.6671  1.2723 -1.1277
#>  -0.2272 -1.9456 -0.3421 -3.2339  3.1313
#>  -1.4981 -1.3352  1.3184 -0.3911  0.6659
#> 
#> (3,.,.) = 
#>  -0.1647  0.0008  0.4195  0.6568 -0.2479
#>   0.2833 -0.3681 -2.1422 -0.1373 -2.4804
#>  -1.0905  0.9223  0.9962 -1.8491  1.1189
#> 
#> (4,.,.) = 
#>   0.1371  0.4425  0.9952  1.8871 -0.2902
#>  -1.2767 -0.4945  0.1557  1.8520 -3.2727
#>   0.3469  0.0902 -0.1055 -1.3771  1.6447
#> 
#> (5,.,.) = 
#>  -0.6114  0.1862 -2.2116 -0.2221 -4.1611
#>  -0.0989 -0.1072 -0.8892 -0.1625 -1.2879
#>   1.2899 -0.6198 -0.7682 -1.6970  2.7560
#> 
#> (6,.,.) = 
#>  -0.9002 -0.6686  0.7476 -0.4892  0.5102
#>   1.2072 -0.2425 -1.8316 -2.6264  1.7271
#>   1.3181 -0.9016  0.5380  0.7148  2.6132
#> 
#> (7,.,.) = 
#>  -1.3253 -2.3322 -0.1508 -3.8124  2.4456
#>  -1.7344  0.4342  1.2367 -0.7748 -0.3066
#>  -0.1834 -0.1359 -0.4975  1.1753 -2.0854
#> 
#> (8,.,.) = 
#>  -3.7992 -2.2777 -0.6481  1.6047 -7.3195
#>   0.1569  0.5818 -0.8449  3.0991 -4.2488
#>   0.3792  0.0464  0.5434  1.5578 -0.0924
#> 
#> (9,.,.) = 
#>  -1.7851 -1.7766  0.4640 -3.3044  1.9087
#>  -1.8305 -1.3658  0.0552 -3.1704  0.8987
#>  -1.4499 -1.3060  0.4094  0.2355 -1.2374
#> 
#> (10,.,.) = 
#>  -0.7771 -0.4226 -0.4226 -2.4339  0.6671
#>  -0.0736  0.0043  1.4790  1.5012  0.6867
#>  -1.5934 -0.1540  0.9720  0.5761 -1.4871
#> [ CPUFloatType{10,3,5} ]