Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> 1.93692
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.2430
#>  0.6136
#> -0.4242
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  2.8780 -3.2541  0.6744
#>  5.0005  3.2526 -2.4646
#> -3.3225  4.0710 -0.1191
#> -1.2687 -1.3322  0.4068
#> -3.5382  0.4271  2.0785
#> -0.6726  0.3660 -0.1558
#>  2.0206  1.4550 -2.4804
#> -2.4879  0.1138 -1.8734
#> -2.1753  3.8180  1.3526
#> -0.1180  0.7907  1.2285
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   1.3677 -0.2370  1.9572 -3.5082 -0.8928
#>  -0.2134 -0.2851 -1.0719 -0.2241 -0.0488
#>  -1.0711 -0.7724 -0.2308 -0.4505  0.4744
#> 
#> (2,.,.) = 
#>   2.2443 -4.9166  0.0623  0.9347  1.3344
#>   2.2231  2.8962 -3.2811 -1.7465  2.0790
#>  -0.1270  0.8903 -0.1488 -0.2443 -0.5845
#> 
#> (3,.,.) = 
#>  -2.0531 -0.2512 -1.6019  1.7093  1.3958
#>   1.2112 -3.0839 -1.2098  0.4753 -2.8606
#>   6.2120  3.5345  2.1577 -8.7470  2.3425
#> 
#> (4,.,.) = 
#>  -1.2379  1.4427 -2.1556  0.1817  3.3683
#>   2.5404 -1.5040  0.7273 -0.5776 -2.1769
#>   2.6874 -3.0415 -1.7356  0.5976  0.6024
#> 
#> (5,.,.) = 
#>  -1.9501 -2.0434  0.3283 -1.0486 -2.4069
#>  -0.8784 -0.2293 -1.1097  0.9006  1.0632
#>  -3.3238  0.4199 -0.5226  0.5094  0.6533
#> 
#> (6,.,.) = 
#>  -0.3981 -1.9177  0.9325 -2.1065 -2.5214
#>  -0.1722  4.9801 -1.7217  1.6814 -0.9674
#>  -0.2151  2.5620  0.9561  1.7791 -1.9273
#> 
#> (7,.,.) = 
#>   0.4792  0.2332  3.6533 -2.0230  5.1757
#>  -0.7508 -1.9095 -0.5521  0.8888 -0.5091
#>  -0.4089 -1.9761 -2.6486  2.3319 -0.9851
#> 
#> (8,.,.) = 
#>   1.2660 -1.0095 -0.8007  0.8927  1.6459
#>   0.9237 -7.7392 -4.5117  2.3187  2.0209
#>  -0.1058  2.1752  1.6319 -1.3814  0.2493
#> 
#> (9,.,.) = 
#>   0.5991 -0.1299 -3.2639 -1.2186  0.6193
#>  -1.3866 -0.0773  1.7161  1.1154  1.4457
#>   0.3977 -0.5484  0.1279  1.0237  0.2254
#> 
#> (10,.,.) = 
#>   1.1244 -0.4213 -2.4000 -2.8574 -2.1364
#>   0.1681  1.9372  0.3854 -1.5693 -2.2065
#>  -1.4761  3.5861  3.8406  0.7081 -0.4562
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -1.3487 -1.3422 -0.1517  2.0762 -0.5363
#>  -3.0505 -2.3564  0.3730 -0.3326 -4.1332
#>  -1.1951 -0.7904  0.6398  0.4018 -3.3959
#> 
#> (2,.,.) = 
#>  -2.8722 -0.4628  0.8062 -3.5852 -5.2110
#>   3.8859  1.2702  0.5281 -0.5280  0.6427
#>  -1.5661  0.0435 -0.0418  0.3425 -0.6706
#> 
#> (3,.,.) = 
#>  -0.1993 -0.1760  0.3780 -2.7035 -1.8777
#>  -2.2992 -1.3876 -0.8823  2.5749  1.7419
#>  -1.1964 -0.5641  0.0031 -1.8218 -1.0861
#> 
#> (4,.,.) = 
#>   3.4628  2.2400 -0.1498 -0.4665  3.3921
#>   1.2075  0.1778  0.0272  0.7030  0.7134
#>  -1.1392 -0.2783 -0.2469 -0.5370  0.1363
#> 
#> (5,.,.) = 
#>  -0.0123  0.7571 -1.0185  0.5191  4.2332
#>  -1.7971 -1.3214  0.2048 -0.2113 -2.3519
#>   1.0772 -0.5960  0.0797 -0.2904  0.0294
#> 
#> (6,.,.) = 
#>  -3.3151 -1.0309 -0.5543  1.3306 -0.0487
#>   1.7564  0.9842  0.5262  0.8817 -0.5317
#>   0.5547  0.8146  0.0670  2.6227  0.6365
#> 
#> (7,.,.) = 
#>   1.0473  0.8361 -0.2914  0.0535  2.0469
#>  -3.5536 -2.5551 -1.2554  4.4593  2.1560
#>   1.8544  0.2943  0.5665 -1.0144 -1.0911
#> 
#> (8,.,.) = 
#>  -0.5588 -0.5715 -0.8905  3.3183  3.1553
#>   0.5509  1.3774 -0.6972 -2.5570  3.2921
#>   3.7443  1.4351 -0.0522  0.6172  2.9493
#> 
#> (9,.,.) = 
#>   2.1801  2.1434 -0.7021  0.1516  4.7941
#>  -0.9632  0.7728 -0.6155 -1.3570  1.9942
#>   2.9432  2.5916 -0.0839 -0.6859  2.9727
#> 
#> (10,.,.) = 
#>   1.8271 -0.0031  0.6844  1.5955 -1.4334
#>  -2.2287 -2.5246  0.0818  0.7835 -2.5267
#>  -2.1468 -0.3054 -0.0360 -0.7511 -1.2646
#> [ CPUFloatType{10,3,5} ]