Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.98052
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.0915
#> -0.4803
#>  0.1216
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.8133  1.1063  1.3090
#>  4.3649 -0.9861 -0.9484
#> -0.7668  0.2129 -0.2296
#> -1.1764 -4.4576 -1.0427
#> -2.1544 -1.7819 -2.3941
#> -4.9113  3.1843  0.5698
#> -1.3242 -0.0767  0.7157
#>  0.2347 -1.2068  3.8221
#>  1.3283 -2.5615 -1.3238
#>  0.3641  0.4202  5.2516
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -1.2165 -2.7346  0.1847  3.1132 -0.9576
#>  -0.1750 -0.1394  0.7393  1.1322 -0.3289
#>   1.6392  0.0277 -0.2499 -0.5371  2.3237
#> 
#> (2,.,.) = 
#>  -0.1097  0.7331 -0.0397 -1.3870  2.8443
#>  -0.2409  0.3011 -0.9890 -0.2164  1.0473
#>  -0.6430  0.2691 -0.7538 -0.3657  1.0435
#> 
#> (3,.,.) = 
#>   0.7113 -1.2512 -0.6192  1.3470  0.0697
#>   0.5567 -0.8821  0.4482  1.1729  0.3508
#>  -1.9476 -2.9420  5.9115 -1.9642 -3.3421
#> 
#> (4,.,.) = 
#>  -2.7279 -2.3519 -3.4962 -1.8454  0.2952
#>  -0.0917  3.2803  2.2842  2.6555  0.6128
#>   1.6096 -0.3718 -1.3668 -1.4122 -0.3896
#> 
#> (5,.,.) = 
#>  -2.7603 -1.3881  1.6565 -2.6275  2.5845
#>   2.5211  2.6449 -0.6835 -1.0770 -0.8359
#>   4.4098  1.5104 -0.0084 -0.5178 -3.6947
#> 
#> (6,.,.) = 
#>  -2.8421 -0.6536  2.3330  1.6759  1.8693
#>  -0.1875  0.4286  1.1292 -0.8807 -0.1597
#>   1.4919  1.0388  1.8531 -4.0056 -3.1985
#> 
#> (7,.,.) = 
#>   1.5208 -0.8174  0.7788 -3.3769  0.4279
#>  -0.4062  1.2204  0.3962  6.0043 -0.2476
#>  -1.1297 -0.8162 -1.1938 -0.6805 -0.2886
#> 
#> (8,.,.) = 
#>  -0.3097 -0.5236  0.0541 -1.3140 -1.8386
#>  -2.3916  1.1753 -0.7198 -0.0499  1.6123
#>  -1.4409  2.0329 -1.0229  3.2992 -0.9494
#> 
#> (9,.,.) = 
#>  -3.9666  2.1826  1.0016 -2.0595 -0.6597
#>   0.9553  0.8955  0.3203 -1.0528 -1.7101
#>  -1.2053  0.0330 -0.1557 -1.2127 -0.5608
#> 
#> (10,.,.) = 
#>   0.3902  0.6571  2.2867 -1.8772  3.0385
#>  -0.6810 -0.2411 -1.0153 -0.0459 -1.0307
#>  -1.6677 -0.2461  1.8351  0.2226  1.6543
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   1.1474  1.9511 -1.0121 -0.0405  1.3895
#>   0.5580 -4.3577  1.3727 -0.9916 -2.2133
#>  -0.4171  1.0962  0.2108 -0.6524  1.5692
#> 
#> (2,.,.) = 
#>  -0.9469 -0.1175  0.8171 -0.7192  0.7886
#>  -2.7153 -6.3253  4.1977 -1.7583 -2.5220
#>  -0.1269 -2.1454  1.3397 -2.5725  1.2167
#> 
#> (3,.,.) = 
#>  -5.3878 -1.7436  4.3569 -1.9035  1.3024
#>   0.9807  2.0425 -1.2561  0.7209  0.6420
#>   7.9363  4.4045 -6.8691  2.4657 -0.1838
#> 
#> (4,.,.) = 
#>  -1.3222 -1.2315  0.8491  0.4263 -1.2562
#>   0.2211  2.1194 -0.5375 -0.7384  2.3603
#>  -0.3109  0.9888 -0.0489 -0.3535  1.1502
#> 
#> (5,.,.) = 
#>  -0.5486  0.9569 -0.2541  1.1549 -0.4924
#>  -0.7094 -2.1146  1.2335 -0.1820 -1.3032
#>  -3.3788 -3.7787  3.1325  0.5616 -3.0915
#> 
#> (6,.,.) = 
#>   3.0435  3.8228 -2.7917 -0.9955  3.6304
#>   2.8293  3.3970 -2.6648 -0.3077  2.5864
#>  -0.7713 -1.4287  1.0548 -0.5711 -0.3656
#> 
#> (7,.,.) = 
#>   1.1784 -0.1683 -0.5254 -0.6180  0.4325
#>   0.2213 -0.9106 -0.0427  0.5210 -1.2674
#>   4.3432  2.8587 -3.4393 -0.5140  2.2615
#> 
#> (8,.,.) = 
#>  -0.4073 -0.9500  0.6260 -0.1885 -0.4608
#>   0.0163 -0.7709 -0.1913  1.4753 -2.1824
#>  -2.6043  1.7585  0.9636  0.3255  1.2075
#> 
#> (9,.,.) = 
#>   1.4516  2.5749 -1.7179  0.4731  1.2603
#>   2.4206  0.9187 -1.9504  0.6515 -0.2627
#>  -2.9325 -2.0715  2.4717 -0.4191 -0.7998
#> 
#> (10,.,.) = 
#>  -2.0689  0.6484  1.0804  0.6741 -0.0458
#>  -2.6179 -3.6950  2.7358  0.5041 -3.0391
#>  -4.0565 -1.1599  2.9709  0.3315 -0.8232
#> [ CPUFloatType{10,3,5} ]