Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> 0.545778
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.3713
#> -2.6815
#> -2.9605
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -2.3016 -1.3684 -0.9073
#>  2.8582 -2.2133 -0.8736
#>  0.9304  2.1700  0.7925
#>  0.6843 -0.5460  1.3034
#> -0.4755 -0.2324 -0.2926
#>  1.5735  0.7525  0.4386
#>  0.9137  1.3856 -0.6690
#>  0.4178 -0.1053  0.3746
#>  2.0224  0.1602  0.1901
#> -1.4167  1.0603 -0.6605
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -1.7619 -3.0383 -0.3035  6.0426  2.0073
#>   2.3486 -1.2095 -1.4159 -0.8657  3.3583
#>   1.8613  2.9206  1.9154 -3.5986 -4.2283
#> 
#> (2,.,.) = 
#>  -1.1387  0.5371  1.1028 -0.0919  0.0345
#>   1.3800 -0.1195 -0.0357  0.3448  0.5509
#>   0.0303 -1.7338 -1.8339  0.6827  1.7457
#> 
#> (3,.,.) = 
#>  -0.2642 -0.5672 -0.4495 -0.3243 -0.3938
#>  -0.9750  0.0324  1.0439 -0.2742  0.0929
#>   2.0733 -5.7320  2.9794  3.3251  2.5726
#> 
#> (4,.,.) = 
#>  -1.5735  0.3575  0.8899  1.1723 -0.9941
#>  -1.5039  0.1245 -4.8765  4.6726 -1.9003
#>   2.7669 -0.9937 -2.8131  0.5230  1.0805
#> 
#> (5,.,.) = 
#>   1.0204  1.0631  1.9662 -1.7794  4.1426
#>  -0.8404  2.8798 -2.7409 -1.2171  4.5267
#>  -0.2625 -2.1690 -2.4959 -1.0591 -3.3227
#> 
#> (6,.,.) = 
#>  -0.6771 -0.8463 -0.0476 -0.8232 -1.3524
#>  -1.9657 -1.8497 -1.6551  0.5963 -4.0864
#>  -0.7453 -2.0213 -1.0328 -0.7417 -2.1930
#> 
#> (7,.,.) = 
#>  -2.3591  2.7839 -1.6736  1.0548 -0.5277
#>  -3.0186  0.8973 -1.5105  0.4855  0.1286
#>   0.2004 -1.2671  0.7929 -1.4683  0.1781
#> 
#> (8,.,.) = 
#>   3.9457  4.4349  4.4247 -1.2236 -0.1888
#>  -1.5154 -2.3542 -0.6079  1.1165  1.6919
#>   0.0076 -1.6295 -2.3815 -0.2497 -0.9117
#> 
#> (9,.,.) = 
#>   1.7494  3.4203 -1.3271 -3.1990 -3.8572
#>  -0.0757  1.4610 -0.4481 -0.7975 -1.1678
#>   1.9226 -1.0074 -1.1155  0.5677  0.6131
#> 
#> (10,.,.) = 
#>   0.3461  0.8091  1.9341  3.5864 -1.2219
#>   0.0496 -2.5217 -1.6121 -0.4448  0.0920
#>  -0.7696  1.2504  2.4948  0.8963 -1.3370
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -2.1746  1.9840 -0.1218  0.4858  0.6265
#>  -0.7202 -1.1964 -2.9350 -1.5753  5.1830
#>   1.1032 -1.0191 -0.6128 -0.6926  1.1545
#> 
#> (2,.,.) = 
#>  -0.0330  3.3266  1.9146  0.6086 -2.3344
#>   0.4839 -1.6629 -1.6087 -0.1589  4.1239
#>   4.1696 -3.0859  2.2124  1.0432 -3.7415
#> 
#> (3,.,.) = 
#>   1.6661  2.7639  3.3926  1.7502 -3.9087
#>   3.8755 -4.7448  0.7102 -0.3763 -3.0279
#>   0.5220 -3.3985 -1.7420 -0.6359  2.1932
#> 
#> (4,.,.) = 
#>  -1.0279  2.1793  1.1775  0.6926 -1.8940
#>  -0.2359 -1.7907 -1.3103 -0.6101  1.4177
#>  -1.1369  1.4732 -0.0336  0.8027  1.5774
#> 
#> (5,.,.) = 
#>  -0.0987 -0.1615 -0.0017 -0.1420 -0.4992
#>   1.6590 -1.8270  0.0749 -0.0315 -0.1180
#>   2.1391 -2.7735  1.2811  0.0939 -4.2653
#> 
#> (6,.,.) = 
#>   0.1709 -1.4468 -0.7134 -0.2098  0.8871
#>  -3.6931  3.6656 -0.4731  0.0427  0.8714
#>   1.0680  0.3984 -0.1897 -0.9971  0.4737
#> 
#> (7,.,.) = 
#>  -1.5284  1.9070  0.5841  0.9527 -0.2711
#>  -2.6205  3.2182  0.0063  1.4593  2.5157
#>   0.8232 -1.6387 -0.1726 -1.0061 -1.7446
#> 
#> (8,.,.) = 
#>  -1.0087  1.2964  1.4189  1.1505 -2.6976
#>  -1.5094 -1.5792 -2.8526 -1.5725  3.6735
#>   4.8443 -3.7456  0.7866 -0.6173 -1.6949
#> 
#> (9,.,.) = 
#>  -2.7305  2.0261 -1.9910 -1.3147  3.1753
#>  -2.0216  1.6484 -0.2920 -0.1845 -0.1012
#>   0.0400  4.1796  3.4262  2.2235 -3.7291
#> 
#> (10,.,.) = 
#>  -1.7930  2.5304  0.8620  1.5289  0.0721
#>   0.8734 -2.3165 -0.6319  0.5089  1.8503
#>  -3.3283  1.9992 -1.3108  0.2040  2.8336
#> [ CPUFloatType{10,3,5} ]