Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.675277
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.2060
#>  0.5844
#>  0.2755
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.0206  5.0464  1.8279
#> -1.2399  1.3700  1.1977
#>  1.8596 -0.0043 -0.6793
#>  3.5538  1.1735 -1.4275
#>  1.6899 -0.5522 -0.8164
#> -4.6748 -0.9192 -0.8258
#> -0.1632  3.6212 -3.6813
#>  0.3355 -0.8424  2.6288
#>  3.0397  2.0644  0.3480
#> -0.5141  1.3984  3.1649
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   0.2980  2.0958 -2.5706 -1.8915  1.9803
#>   0.8761  0.0544  0.1502  1.0405 -0.0349
#>   3.2095  1.3180 -1.4217  1.1342 -0.4610
#> 
#> (2,.,.) = 
#>   0.2454  3.5178 -3.5613  0.8913  1.9016
#>  -0.8665  5.3806 -2.9317  1.7177  2.5128
#>   0.7254 -1.2062 -0.9617 -0.5736 -0.4082
#> 
#> (3,.,.) = 
#>   4.3358 -1.2212  0.0134 -5.6731  2.3795
#>  -1.7737 -1.8709 -0.6432  3.9838 -6.1614
#>  -4.4704  2.2107  1.2994  3.6471  1.6229
#> 
#> (4,.,.) = 
#>   0.0849 -3.8606 -1.2789 -0.5455 -0.4480
#>   0.9023 -0.7174  0.0925  0.3400 -0.2140
#>   2.7504  3.6400  1.8898  1.7600 -0.7363
#> 
#> (5,.,.) = 
#>  -0.5975  1.4401 -1.9608 -0.3719 -0.6569
#>  -0.6743  1.7375 -1.9964  1.9847 -0.7352
#>   0.9057 -0.9645 -0.2283  0.7341  0.6869
#> 
#> (6,.,.) = 
#>   2.1970 -1.1575 -0.6109 -2.4511  0.0380
#>  -0.8343  0.7137  0.9173  1.9673 -0.5383
#>   0.2945  0.2069 -0.1117 -1.4676 -1.5787
#> 
#> (7,.,.) = 
#>   1.2318 -1.7093  0.8489  1.9776 -7.4257
#>  -0.1868  1.1615 -0.5042 -1.3579  0.8986
#>   0.1639 -0.0301  0.2196  0.3213  0.3607
#> 
#> (8,.,.) = 
#>  -0.5042 -0.1776  1.7233 -1.8370 -2.7092
#>   1.3716  0.6140 -2.1166  0.6911  2.5284
#>  -3.9501 -0.1454  3.4784 -3.9339 -6.1042
#> 
#> (9,.,.) = 
#>   5.9626 -0.9115 -1.4580 -3.4294 -4.5488
#>   1.8878 -1.6196  1.9431 -0.3934 -0.8121
#>   2.5907  0.2763 -0.3963 -0.8267 -1.1802
#> 
#> (10,.,.) = 
#>   0.7538 -1.6682  0.5534  0.5846  2.9602
#>   0.4702 -6.9517 -0.3360  1.7368  6.0791
#>   1.0323 -2.9231 -0.5006 -0.2553  2.7502
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.8752 -1.0135  0.3928 -0.2168 -0.1929
#>   1.5095  0.9876 -1.8189  0.2087  0.3454
#>  -1.6264 -0.9879 -2.1097 -0.0604  0.4966
#> 
#> (2,.,.) = 
#>   3.2208  0.2826  0.7164 -0.3108  0.8912
#>   6.9691 -1.3032  0.2221 -1.1515  1.7788
#>   0.7156 -0.7033  2.2447 -0.3885  0.1731
#> 
#> (3,.,.) = 
#>   0.3010 -0.4362  1.0364 -0.1844 -0.3191
#>  -5.0082 -2.8808 -3.2293 -0.1547 -0.7954
#>   0.3525 -2.4974  3.2374 -0.8865 -0.3615
#> 
#> (4,.,.) = 
#>   2.4745  3.0972  2.7333  0.5337 -0.0911
#>  -3.0236  1.5341 -0.5707  0.8197 -1.0685
#>  -1.8825  1.7039 -0.4994  0.7410 -0.7959
#> 
#> (5,.,.) = 
#>  -1.2167  0.2504 -0.3692  0.2314 -0.4395
#>  -1.6841 -0.2087 -1.1345  0.1754 -0.3658
#>  -0.0109 -1.8764 -0.6022 -0.5228  0.1951
#> 
#> (6,.,.) = 
#>   3.3188  1.2600  3.6204 -0.1215 -0.0335
#>   4.6209  2.9625  1.0463  0.3584  0.1594
#>  -4.4356  1.2109 -1.4533  0.8629 -0.4264
#> 
#> (7,.,.) = 
#>   6.6341 -0.1328  0.8080 -0.8086  1.7424
#>   1.6623 -2.1509  3.0035 -0.9653  0.7284
#>  -0.6211  0.4913 -0.1247  0.2165 -0.1709
#> 
#> (8,.,.) = 
#>   1.8104 -3.4859 -0.8713 -1.1934  0.9830
#>  -1.9554  0.3437 -2.2523  0.3777  0.2274
#>   1.7286 -4.8184  2.1428 -1.7095  0.9009
#> 
#> (9,.,.) = 
#>  -0.3630  1.3626 -1.6407  0.5198 -0.2308
#>  -5.4670 -0.3200 -3.3395  0.6279 -0.7289
#>   0.8029 -2.7253  0.5438 -0.8720 -0.2502
#> 
#> (10,.,.) = 
#>   0.3447  0.1062 -1.2406  0.0727 -0.2450
#>   1.1545  0.7022 -1.1217  0.1222  0.4148
#>  -0.6421 -0.7258  2.6897 -0.2503 -0.4523
#> [ CPUFloatType{10,3,5} ]