Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   2.6101  3.5925 -1.5538  0.9263  6.4930
#>  -2.4585 -2.9103  1.6708 -0.6435 -5.7488
#>   1.5055  1.2162 -1.3177  0.7266  2.8734
#> 
#> (2,.,.) = 
#>   1.8401  1.0012  0.5018 -0.9943  2.0108
#>   0.1859 -1.4439  1.3300 -1.4212 -2.2684
#>   2.4771  2.1502 -1.5628  0.3589  4.7941
#> 
#> (3,.,.) = 
#>  -2.3101 -1.0256  0.6809  0.7654 -3.0363
#>  -0.4339 -2.0531  1.1157 -1.0809 -3.2684
#>  -3.1691 -3.8237  2.0672 -0.8540 -7.4284
#> 
#> (4,.,.) = 
#>  -1.4519 -3.4693  1.1928 -1.0652 -5.5652
#>   0.5104 -2.2056 -0.1386 -1.0967 -2.1429
#>  -1.3685  0.5733  0.7915  0.1328 -0.4249
#> 
#> (5,.,.) = 
#>   0.6131  1.6648 -1.6860  2.3184  2.7681
#>  -2.0384 -2.5822  2.0773 -1.5012 -5.1302
#>  -1.8800 -2.0524  0.2205  0.4939 -3.7535
#> 
#> (6,.,.) = 
#>  -1.5723  0.2513 -0.5902  1.8205 -0.5387
#>  -1.9894 -3.4887  1.9805 -0.8728 -6.4752
#>  -0.7823 -1.1439 -0.0047  0.0258 -1.7941
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]