Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   0.1238  1.5913  1.6534 -1.6017 -2.2940
#>  -2.1857 -1.6468  0.2380  2.9087  1.8575
#>   0.9375 -0.2953  0.6460 -0.9955 -0.0939
#> 
#> (2,.,.) = 
#>  -1.6683 -0.6259  2.4767  1.8254  0.4255
#>   0.3493 -1.6113  0.3988 -1.2335  0.2699
#>   0.0064  1.7142  2.5851 -0.3798 -1.6961
#> 
#> (3,.,.) = 
#>   0.8092  1.6601  0.7839 -2.7390 -2.6467
#>  -0.6973 -1.7651  6.3127 -1.9308 -1.5240
#>   0.9800 -1.0079 -1.2720 -2.3656 -0.3333
#> 
#> (4,.,.) = 
#>   1.2242  0.4933 -0.5252 -1.9969 -0.9790
#>  -1.0172 -1.8660  3.6142  0.4252  0.4092
#>   1.5987  0.5170  3.5435 -2.5458 -1.7388
#> 
#> (5,.,.) = 
#>   2.0124 -0.6038  4.2165 -4.8549 -2.5426
#>   1.1765  0.9191  2.7707 -3.6168 -2.8781
#>   0.8835 -0.9041  1.1681 -0.6078  0.4608
#> 
#> (6,.,.) = 
#>  -1.0249 -0.2399 -1.7317  1.5067  0.8938
#>   0.1071 -2.1920  5.5769 -1.2728 -0.2433
#>   0.7367  2.3370  0.1251 -1.6494 -2.2452
#> 
#> (7,.,.) = 
#>  -1.6529 -0.6533 -1.3089  2.8275  1.6858
#>  -0.8665 -1.9948  2.4356  0.0284  0.4538
#>  -0.7499 -1.1948 -9.2200  2.1732  3.1508
#> 
#> (8,.,.) = 
#>  -0.1857 -2.5828  1.4622 -0.4958  0.9597
#>   0.5854  1.5191 -1.4361 -0.3276 -0.6620
#>   0.6868  1.2184  1.1130 -1.8035 -1.8185
#> 
#> (9,.,.) = 
#>   0.6339  1.8226  5.8859 -1.9212 -2.9684
#>  -3.5368 -1.4109 -1.4991  7.1448  4.2137
#>  -0.4163  1.6408 -1.3197  0.9628 -0.3906
#> 
#> (10,.,.) = 
#>   0.0257 -1.4893  1.9282 -0.1886  0.5461
#>   1.2470  0.0136 -3.6270 -0.1386  1.0825
#>  -0.4410 -0.6210  4.1224  0.4227 -0.1237
#> [ CPUFloatType{10,3,5} ]