Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.4744 -1.8977 -1.0119  0.6278 -1.8236
#>  -1.4078 -3.1783 -2.9031 -0.7162  0.1102
#>  -0.2563 -2.2930 -2.5515 -0.9151  1.7648
#> 
#> (2,.,.) = 
#>   0.0180  1.4048  0.5549  0.4986 -1.4930
#>   2.0518  2.9535  2.1656  0.8557  1.0933
#>   2.7551  7.3454  3.5064 -1.4805  0.5484
#> 
#> (3,.,.) = 
#>  -0.7438  2.0755  4.6408  2.5857 -4.3768
#>  -2.4249 -3.7304 -2.1264  0.4577 -2.1070
#>  -0.6990 -3.0698  2.7345  1.4971  0.7726
#> 
#> (4,.,.) = 
#>  -2.4552 -6.1879 -1.2140  2.0668 -1.0144
#>  -1.4039 -2.7783 -1.3367 -0.3735 -0.2044
#>   0.6912  1.5408 -0.8640 -0.8645  0.4523
#> 
#> (5,.,.) = 
#>  -1.3195 -1.2740  2.8774  2.6049 -2.8441
#>  -1.6668 -2.9876 -0.3207  0.5131 -1.0053
#>   0.6981  0.5015 -0.9685 -0.3043  0.9514
#> 
#> (6,.,.) = 
#>   1.3506  3.4776  3.0382  0.4207 -0.2396
#>   1.9470  4.6467  1.2540 -1.4079  0.9172
#>  -0.2607 -1.0281 -0.1930  0.0310  0.3542
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]