Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   1.2454  0.9743  0.8746 -0.0979 -3.8157
#>   2.0364 -0.3706  0.1288  3.5717  1.8004
#>   3.6495 -0.5521  1.9564 -2.1824 -2.6201
#> 
#> (2,.,.) = 
#>  -0.5158  0.4049 -0.3542 -0.5950  1.9541
#>   1.4718  0.5877  0.1496  3.0827 -0.0265
#>   8.2489 -1.5364  3.6701 -2.4705 -1.8861
#> 
#> (3,.,.) = 
#>  -1.5114 -0.9782 -0.5966 -1.1418  1.8374
#>   2.4624 -0.9457  1.0457  1.1674 -2.0717
#>   0.0595  1.0234  0.8161 -3.3060 -3.7096
#> 
#> (4,.,.) = 
#>   4.3079 -0.1204  1.2879  1.6750  1.0951
#>  -3.5555  1.2728 -1.1073 -2.5166  0.3693
#>   1.6589  0.2245  0.6002  0.4822 -0.4093
#> 
#> (5,.,.) = 
#>   5.5507 -1.1722  2.6225 -2.5551 -1.5937
#>  -3.3346 -1.0664 -0.9903 -1.8894  0.1194
#>   0.6886 -0.3522  0.1091  1.0541  0.2698
#> 
#> (6,.,.) = 
#>   2.8779 -0.3767  1.2373 -0.8873 -0.2979
#>   0.5591 -1.1803  0.4697 -2.4665  0.7981
#>   1.3783 -0.6951  0.6482 -1.3374  0.5367
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]