Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -1.0854 -0.3106 -1.0033 -0.3564  0.1341
#>   3.0233 -3.5604  0.4568 -2.0943 -2.3565
#>  -0.2096 -1.1447 -2.3411  1.6757 -0.4973
#> 
#> (2,.,.) = 
#>  -2.9318  1.9395 -0.7596  0.0682  1.5585
#>  -0.4162  1.0273  1.0468 -1.7403  0.1172
#>  -0.6692  2.9135 -0.6570  3.3464  0.8916
#> 
#> (3,.,.) = 
#>  -1.5417  1.5248 -1.4330  2.0953  0.8615
#>  -3.1345  2.8064 -1.2249  1.7878  1.9470
#>  -1.0756 -0.0637 -0.4569 -1.2669  0.0654
#> 
#> (4,.,.) = 
#>   0.1776 -1.8026 -2.0886  2.7294 -0.0511
#>   1.4663 -0.6895  0.2878  0.6218 -0.6258
#>   2.0762 -0.8575 -0.1884  1.5545 -1.0435
#> 
#> (5,.,.) = 
#>  -3.5785  4.5006  0.1386 -0.0867  2.1047
#>   0.8325 -0.4014 -0.9571  1.1381 -0.8058
#>   1.1124 -1.3057  1.5762 -2.4621 -0.6010
#> 
#> (6,.,.) = 
#>  -1.8304  2.6743  0.1596 -0.2852  0.9528
#>  -0.7088 -0.9934 -0.2234 -0.3646  0.5113
#>   0.6915  0.7325 -0.0696 -0.1990 -0.8863
#> 
#> (7,.,.) = 
#>  -0.0633 -0.5789  0.4502 -1.4199 -0.1084
#>   5.4557 -2.9810  3.7907 -2.3379 -2.3132
#>   2.9668 -3.2792  0.1192 -0.0693 -1.7412
#> 
#> (8,.,.) = 
#>   0.7537  0.0883  2.0963 -1.6989  0.0870
#>   1.0683 -0.5934 -1.3138  3.0819 -0.3823
#>  -2.8278  1.5437 -0.5967 -1.6975  0.9164
#> 
#> (9,.,.) = 
#>   1.4877 -0.8614  0.1503  0.9179 -0.5757
#>  -2.5814  2.8638  0.5181 -0.0891  1.8865
#>   2.0143 -0.3321  0.1814  1.5253 -0.8599
#> 
#> (10,.,.) = 
#>   0.0128  0.4085 -0.3478  1.8907  0.4632
#>  -0.9882 -1.1298 -2.5092  0.2353 -0.5731
#>   0.5268  2.7097  3.7926 -2.1421  0.7009
#> [ CPUFloatType{10,3,5} ]