Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   2.8647  2.3414 -2.5318  0.2337 -3.8611
#>   2.6068  0.7724 -2.7038 -0.8900 -2.2710
#>   2.2201  1.2841 -3.7193 -4.3678 -4.2706
#> 
#> (2,.,.) = 
#>   2.7699  1.0000 -2.9149 -1.2348 -2.0660
#>  -1.9605 -3.1423  2.1686  0.8500  5.1197
#>   1.5428  4.3262 -3.0670 -4.3622 -6.8764
#> 
#> (3,.,.) = 
#>  -1.9904  0.5224  0.9883 -2.3249  0.1253
#>   0.3398  5.2984 -1.2097 -2.4111 -6.4814
#>  -3.1826 -0.0298  2.2785 -1.5211  0.6012
#> 
#> (4,.,.) = 
#>   0.2546 -1.3613 -0.0485  0.8398  0.3466
#>  -0.0357 -1.9669 -0.0395 -0.2414  2.2114
#>   0.3113  0.2481  0.8796  2.7450  1.6671
#> 
#> (5,.,.) = 
#>  -2.4251  0.1399  2.5915  0.7187  2.2659
#>   3.2235  3.7672 -4.2524 -3.3423 -7.1123
#>  -1.2216 -1.4834  2.4775  3.5517  3.5230
#> 
#> (6,.,.) = 
#>  -3.1441 -4.0578  5.0317  5.3421  8.9368
#>  -2.1684 -0.6256  1.4305 -1.1948  0.4198
#>  -2.8954 -2.0543  3.5882  2.3262  4.9789
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]