Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   0.8417 -3.0647 -1.5891 -1.8937  0.6794
#>  -1.2890  2.2730 -1.0616  0.8649 -1.5458
#>   0.2863  1.9487  0.5221  1.0935 -1.5940
#> 
#> (2,.,.) = 
#>   3.0913  0.7223  1.3888 -0.1952 -3.2007
#>  -1.0287  3.1796  4.4294  1.4898  0.7586
#>  -0.5114  0.0886  1.0346  0.1153  1.0774
#> 
#> (3,.,.) = 
#>  -1.5049 -3.6475 -1.9584 -1.5683  3.6755
#>   0.3817 -1.9465 -1.4412 -0.1561  1.0421
#>   0.2492  0.7451  1.5996  0.2085 -0.0645
#> 
#> (4,.,.) = 
#>  -0.3072 -0.0932  0.6875  0.2013  0.9547
#>   1.4361  1.2129 -0.6392  0.3016 -3.0423
#>   1.0100 -0.1060  0.0214  0.1983 -0.7048
#> 
#> (5,.,.) = 
#>  -1.3174 -0.6257 -0.3095 -0.0364  1.8352
#>   2.2447  0.6012  0.7079 -0.3355 -2.7270
#>   1.3929  1.5550  1.7449  0.0877 -2.1593
#> 
#> (6,.,.) = 
#>  -0.3755 -0.4950  2.3526 -0.1978  2.2173
#>   0.8485 -1.6378 -1.9336 -0.7724 -0.4991
#>  -0.7703  1.2797 -0.2449  0.8622 -0.3624
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]