Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.7519 -0.0182  0.4102  0.3720 -0.0695
#>  -1.1239  0.5004 -0.1953  0.1638  1.1672
#>  -2.7116  0.1027  1.0568 -0.3926  1.3563
#> 
#> (2,.,.) = 
#>   0.9339  0.3337 -1.0397  0.2413  0.0387
#>  -1.8614 -0.4065  0.7794 -2.5920  1.5129
#>   5.5878  1.5946 -2.1503  1.3727  1.8627
#> 
#> (3,.,.) = 
#>  -1.4311 -0.2688 -0.1510 -1.8608  0.6801
#>  -0.7878 -0.7221 -0.4008  1.2930 -3.3457
#>  -1.3427  0.1291  1.9894  2.6229 -0.4281
#> 
#> (4,.,.) = 
#>  -2.1646 -0.6938  0.6059 -2.0751  0.1701
#>  -1.2848  0.8783  0.7190  0.7926  2.3324
#>   4.0216 -0.2874 -4.3364 -6.8707  1.8385
#> 
#> (5,.,.) = 
#>   1.1277 -0.5353 -2.4101 -9.2005  4.3648
#>   2.6710  0.3919 -1.5916 -1.1608  0.9507
#>  -2.9545 -0.4284  1.7984 -0.3234  0.2977
#> 
#> (6,.,.) = 
#>   4.4994  1.3286 -1.6452  3.5165 -0.2814
#>   0.4981 -1.0431  0.3176  0.8394 -3.4252
#>   0.5387 -0.3335 -1.2571 -3.3342  0.9306
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]