Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.6166 -1.8560 -5.1044 -2.6185  0.6091
#>  -0.4009  1.2607  4.1392  2.8008 -1.8729
#>   1.6467 -0.6901  0.8373 -4.7633  1.7823
#> 
#> (2,.,.) = 
#>   2.4121  3.0835  8.1051  0.6456  0.9470
#>   0.5798 -0.5382 -4.4372 -1.7572  2.4067
#>   0.4184 -0.0288 -4.1735 -0.7685  2.1689
#> 
#> (3,.,.) = 
#>   0.0187 -0.8238 -2.2639 -1.7094  0.6787
#>   1.5292  2.1939  3.0536  0.4526  1.4009
#>   1.0373  2.6470  2.0528  2.7709  1.2592
#> 
#> (4,.,.) = 
#>   2.4733  1.9234  2.8785 -1.5537  2.7554
#>  -0.6725 -1.3798  0.8100  0.1919 -1.3972
#>   1.2046  2.3351  0.8983  1.3330  1.7978
#> 
#> (5,.,.) = 
#>  -3.1460 -0.0672 -2.4116  5.9009 -3.4866
#>   0.4041  1.2139  1.5466  2.0409  0.3194
#>  -0.5355 -1.6918 -1.5472 -2.4975 -0.5922
#> 
#> (6,.,.) = 
#>  -1.0515 -0.1826 -6.2704  2.2816  0.9382
#>  -0.0159 -1.0528 -4.2842 -1.8508  1.3684
#>  -0.9217 -0.5439 -2.9411  0.5836 -0.3141
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]