Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.5919  0.9296  0.9139  0.3764 -0.0398
#>  -0.4627  3.2801 -2.7602 -1.0476  2.5434
#>   1.3699 -3.0746 -0.3935 -0.6062  0.6819
#> 
#> (2,.,.) = 
#>   0.3349  0.2050 -1.0875 -0.9076  2.5842
#>  -0.8871  2.0213  0.4434 -0.0203  1.4912
#>  -0.5640  0.8310  0.9363 -0.2621  2.4895
#> 
#> (3,.,.) = 
#>   0.3331 -2.6183  2.2489  0.2368  0.3957
#>  -0.1864 -1.6350  3.7742 -0.0838  4.0876
#>   0.2556 -0.6238  0.5127 -0.3674  1.9108
#> 
#> (4,.,.) = 
#>  -1.5200  2.8272  1.7239  0.8461 -0.1345
#>  -0.0149 -0.3521  2.2247 -0.7954  5.9518
#>  -0.1745  0.8714  0.1371 -0.4416  2.4810
#> 
#> (5,.,.) = 
#>   0.4963 -1.0149 -2.0749  0.1796 -3.9305
#>   0.7008 -2.6886  1.6146  0.7693 -2.3362
#>   0.0906  0.5243 -4.0617  0.4451 -6.8797
#> 
#> (6,.,.) = 
#>  -1.1527  3.1203  0.1074  0.3624  0.1251
#>   0.6516 -3.4972  4.1380  0.5798  1.2190
#>  -1.2885  4.8913 -0.9176 -0.6882  4.1768
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]