Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   1.8516  0.7327  0.7557  0.3976  3.8611
#>  -1.1578  0.2752  1.7058 -0.3034 -3.4784
#>   0.5338 -0.3243 -1.4301 -0.2846  2.6637
#> 
#> (2,.,.) = 
#>  -0.0095 -0.3185 -0.7954 -0.0466 -0.5671
#>  -0.9172 -0.8467 -1.8524 -0.1815 -0.8911
#>  -1.2171 -0.2913 -0.6040  1.0557  1.0854
#> 
#> (3,.,.) = 
#>  -2.2259 -0.5668 -0.1656 -0.5754 -3.6368
#>  -3.1644 -0.9105 -0.6823 -0.0236 -4.6437
#>  -0.5237 -0.5074 -1.2119  0.1332 -0.0487
#> 
#> (4,.,.) = 
#>  -0.2789 -0.4134 -0.7804 -0.2159 -1.7218
#>  -1.9864 -0.5482 -0.6249 -0.1567 -0.8752
#>   1.8638  0.8441  1.1970  0.0771  3.0981
#> 
#> (5,.,.) = 
#>  -1.1351 -0.3190 -0.4147  0.8268 -0.9241
#>   0.8818 -1.0237 -3.4038 -0.0726  1.8077
#>   2.3053  0.8571  0.8038  0.4822  4.7028
#> 
#> (6,.,.) = 
#>  -0.9657 -0.6691 -1.5390  0.4811  0.0292
#>  -1.1829  0.0124  0.9177 -0.3714 -2.9001
#>  -0.0097  0.0304  0.0866 -0.6854  0.5567
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]