Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   2.5004  0.8594 -1.4965  3.2693  1.8990
#>   2.8841  0.2767 -0.7922 -1.7340  1.4116
#>  -0.4607  1.1174  0.4896  1.4968 -0.8871
#> 
#> (2,.,.) = 
#>  -0.7479 -1.9512  0.5445 -5.1500 -0.5065
#>   1.0755 -0.4039 -1.0811  1.0871  1.9074
#>   1.0745  0.3787 -0.4451  1.0264  0.0126
#> 
#> (3,.,.) = 
#>  -0.9513 -0.2782  0.1543  0.2587  0.1488
#>  -0.2735 -1.5393  0.2950 -4.2482 -0.0143
#>  -0.0085  0.0993  0.2448 -0.5954 -0.5122
#> 
#> (4,.,.) = 
#>   0.1915  1.1129  0.6863 -0.2262 -1.4723
#>   4.6048  1.1699 -1.2027 -0.7502  1.1773
#>  -2.1115 -0.3216  1.0239 -1.3764 -0.6698
#> 
#> (5,.,.) = 
#>  -0.5480 -0.6083  0.5543 -2.5656 -0.6320
#>  -3.7384 -1.2945  1.2670 -1.0439 -1.6679
#>  -4.2150 -0.8546  1.6615 -0.9611 -1.9237
#> 
#> (6,.,.) = 
#>   4.9592  1.7142 -1.8066  2.3538  1.6778
#>   0.7361  0.7982 -0.0480  0.8886 -0.3648
#>   0.8588 -0.0305 -0.4976  0.3691  0.6096
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]