Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   1.5488 -1.0328  0.8628  3.3859 -0.0609
#>   0.0608 -1.1548 -1.9476  0.4227  1.0363
#>  -0.3554  0.4112 -1.9306 -0.0841 -1.8673
#> 
#> (2,.,.) = 
#>  -1.2502  0.3135 -0.4534  1.6507 -0.6913
#>  -0.4289 -1.0374 -2.1931  0.0982  1.1914
#>   0.5608 -0.7919 -1.6293  2.1473 -0.8205
#> 
#> (3,.,.) = 
#>  -0.5661  1.2779  2.3933 -1.0444 -0.2765
#>   2.8929 -1.3530 -1.4997  1.1872 -1.0665
#>   1.8600 -1.4127 -0.8673  1.4477  0.2759
#> 
#> (4,.,.) = 
#>   1.4137 -1.0855  1.1001  1.2062  1.3659
#>   4.7763 -3.2273 -0.5910  3.7120  0.9085
#>   0.2395  0.4844  0.4503 -1.2468 -0.4076
#> 
#> (5,.,.) = 
#>   0.2118  0.0214 -0.3753  0.0350 -0.5023
#>   0.3854 -0.1067  2.2182 -1.0296  1.8759
#>  -0.6757  0.3743  2.4388  0.3935  1.2017
#> 
#> (6,.,.) = 
#>   0.2687 -0.5789  0.2733  2.4855  0.0734
#>   2.5875 -2.7074 -1.2472  1.5518  2.2434
#>  -0.0774 -1.2602 -2.7664  0.6407  0.7424
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]