Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.4258  1.8668  1.1054 -1.5046  0.3015
#>   1.8081 -1.2894 -2.3017 -1.4099 -1.1545
#>  -2.1191 -0.3596  2.5118  0.7671 -3.4207
#> 
#> (2,.,.) = 
#>   1.3663 -2.5897 -3.3939  0.8825  0.9329
#>  -1.2454 -0.2373 -1.4665 -0.3015 -1.2888
#>  -2.5198  0.5830  2.1645  1.7926 -0.1062
#> 
#> (3,.,.) = 
#>   0.0276  0.0635 -0.0724  0.0971  0.3693
#>   3.1424  1.8143  0.9586 -2.0939  2.2718
#>  -1.3494 -1.5718  2.4448  1.9292 -2.5923
#> 
#> (4,.,.) = 
#>  -1.3767  0.4642 -0.2744 -0.5030 -1.4054
#>  -8.1372 -1.8606  1.6029  4.4299 -5.5535
#>  -1.2411  0.4734 -2.2075 -0.4950  0.1643
#> 
#> (5,.,.) = 
#>  -5.4986 -0.6610 -2.7788  2.4936 -0.8137
#>   1.1279  0.5386 -3.1624 -0.1417  4.2257
#>  -1.9528 -1.3417 -0.5582  0.3644 -3.4959
#> 
#> (6,.,.) = 
#>  -0.5713  0.9937  1.8104 -0.5602 -1.1717
#>  -1.7657  0.6612  0.6798  0.4122 -0.5345
#>  -1.4113 -1.3080 -0.3600  2.9276  1.6094
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]