Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.7063 -2.8158 -3.3784 -1.2453  1.6749
#>   0.9368  4.0279  0.3166 -1.4043  0.8740
#>   0.4375  2.1088 -2.4117 -2.5250  2.3136
#> 
#> (2,.,.) = 
#>   0.3300  3.5704  2.5474  1.9546 -1.1200
#>   0.0351  0.5745 -0.6872 -0.4245  0.6567
#>   0.1947 -3.4223  0.0930 -1.2740 -0.2809
#> 
#> (3,.,.) = 
#>   1.1869  4.6685  5.0449  2.0940 -1.7514
#>   0.3382  0.0541  1.0453  0.0378 -0.2990
#>  -0.3640 -3.3143 -3.0398 -2.3340  1.2820
#> 
#> (4,.,.) = 
#>   0.4537  0.2658  0.4264 -0.7947  0.1421
#>  -2.3461 -4.2053 -2.0342  3.3442 -1.4347
#>   0.5724  1.0566  0.2762 -0.9296  0.5699
#> 
#> (5,.,.) = 
#>   3.1720  6.6086  5.7696 -1.8945 -0.0093
#>   0.3171 -0.6962  0.8178 -0.1156 -0.0581
#>   1.0297  4.1702  2.8131  0.4554 -0.6221
#> 
#> (6,.,.) = 
#>  -1.2688 -2.6915 -4.2594 -0.8019  1.3320
#>   0.7898  2.8465  1.1796 -0.2495  0.4743
#>   0.6549  1.4728  1.8807  0.3093 -0.3045
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]