Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   4.5268 -2.2912  2.4896 -0.2483  1.5931
#>   0.1119  1.4485  3.1861  5.1863  3.0936
#>   6.3413 -1.2122 -2.2457 -7.6519 -1.6566
#> 
#> (2,.,.) = 
#>  -3.9824  3.1146 -4.2682 -2.4299 -2.2023
#>   0.7955  1.0566 -2.9583 -4.3873 -1.5295
#>   3.4696  1.4134 -3.0999 -5.8052 -1.9297
#> 
#> (3,.,.) = 
#>  -4.3655 -1.2261  0.3422  3.0909 -1.9555
#>   3.6528  1.5447 -1.3108 -4.8581  3.8001
#>  -1.3947 -0.5520 -1.0263 -0.8146 -1.3385
#> 
#> (4,.,.) = 
#>   3.5468 -1.1617  0.0461 -2.4131 -0.6504
#>   2.1558  0.7507 -2.6914 -4.9873 -1.2659
#>  -1.1141 -1.5001 -3.2813 -4.3494 -4.3236
#> 
#> (5,.,.) = 
#>   0.3025  1.7222  1.4012  1.8228  3.9543
#>  -2.0375 -0.8741 -0.2611  1.0546 -2.3205
#>  -1.5181 -2.1688  2.0160  3.3310 -0.7094
#> 
#> (6,.,.) = 
#>  -2.9451  0.3827 -2.1636 -0.1346 -4.5682
#>   2.0171  0.8700  0.6474 -1.1092  4.2940
#>  -2.5329 -0.0886  2.7099  5.3810  1.9461
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]