Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.1397  0.5486  1.4352 -1.3379 -0.3739
#>  -0.1496  0.1645 -0.1737  0.7425  0.1997
#>   1.3190 -0.5844 -3.1586  1.4744  0.7619
#> 
#> (2,.,.) = 
#>   2.0372  0.5065 -0.0883 -1.7247 -0.3752
#>   0.6848 -0.7902 -1.8468  0.6368  0.0640
#>  -0.9706  0.7548  1.7644 -0.8559  0.3350
#> 
#> (3,.,.) = 
#>   2.4984 -1.0024 -3.8528  2.7208 -0.6444
#>   1.2305 -1.2568 -1.7779  2.2315 -1.7845
#>   0.8073  0.6315 -0.3928  1.6606 -0.2597
#> 
#> (4,.,.) = 
#>   0.9461 -0.2813 -0.5038 -0.0480 -0.6584
#>   1.1584  0.5195 -0.8855 -0.2171  0.6229
#>   1.1611 -0.5093 -2.0081  1.9406 -0.4197
#> 
#> (5,.,.) = 
#>   0.4453 -0.8515  0.5567  0.1493 -2.1192
#>  -0.8998  0.0449  2.5332 -3.3858 -0.3188
#>   3.3887 -1.8495 -4.8165  3.6421 -1.8956
#> 
#> (6,.,.) = 
#>   1.5909 -4.2585 -2.7245  1.2863 -4.4389
#>  -0.6958  1.3803  1.3670 -0.6994  1.2153
#>  -3.3602 -0.8828  1.6869 -0.4401  0.3697
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]