Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.5885  1.6407 -0.9444  1.9821 -1.4706
#>  -1.8012  0.3457 -0.2557  0.3355  0.8735
#>  -1.5355  0.9151  1.2689 -1.5912  0.1659
#> 
#> (2,.,.) = 
#>   2.6086 -2.5154  2.7557 -1.5593 -0.9063
#>  -3.2792  4.3311 -0.5384  0.4077 -0.6023
#>  -0.3389  1.5723 -0.1544  1.1152 -1.0260
#> 
#> (3,.,.) = 
#>  -0.8818  2.2317  1.4916 -0.0828 -1.5919
#>  -0.3025 -0.6109 -1.9682  1.5514  1.0903
#>   0.3625  1.5085 -0.9150  2.5039 -1.4256
#> 
#> (4,.,.) = 
#>   1.0267  0.5690  0.5203  0.1274 -1.2554
#>  -1.9276  1.7339  1.9219 -2.4047 -0.1992
#>   1.3618 -0.9349  1.4153 -2.6563 -0.1493
#> 
#> (5,.,.) = 
#>  -1.8250  1.8908 -2.0561  3.3611 -0.0551
#>  -0.0565 -3.0769  0.7477 -3.2270  2.5149
#>  -1.4354  2.5288  0.9750 -0.2278 -1.1465
#> 
#> (6,.,.) = 
#>  -0.8530  3.3288 -3.0279  4.3328 -1.3121
#>  -2.1757  2.8098 -2.1230  2.2553 -0.0561
#>  -2.7267 -3.4844  3.8681 -7.8286  4.1216
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]