Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -4.8773  0.1019  1.0692 -0.3032 -0.0567
#>   0.1474  2.7347 -1.3818  0.0938  0.1801
#>  -0.2913 -0.3560  1.0115  4.4241 -5.3064
#> 
#> (2,.,.) = 
#>   2.8110  1.6576 -4.9534 -3.8710 -0.6840
#>   2.6988 -2.4793 -0.7497 -1.5179 -0.4120
#>  -2.3990 -1.8612  4.1422  3.5326 -0.4429
#> 
#> (3,.,.) = 
#>  -3.3865  0.0303  2.1729  2.3912 -1.1640
#>   3.2541 -0.8629  1.5117  4.5590 -3.0081
#>   0.9179  0.7346 -0.6374 -0.5204  0.8519
#> 
#> (4,.,.) = 
#>  -2.0744  2.4077 -4.2113 -6.6965  3.1927
#>   6.1996 -1.6839 -1.0973 -0.1304 -0.3404
#>  -2.8564 -0.1204  1.7804  1.8793 -1.0407
#> 
#> (5,.,.) = 
#>  -0.2903 -0.3229  0.8704  1.9286 -1.7180
#>   6.2203 -5.0159  1.4872 -0.6486  2.0183
#>   0.1225  0.8853 -0.2598 -0.4390  1.2068
#> 
#> (6,.,.) = 
#>  -1.6094 -0.3524  0.8585  0.4966 -0.3355
#>  -5.6289  2.1260 -0.0602 -3.3338  4.1298
#>  -0.0214  4.3573 -4.5962 -6.8540  6.0199
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]