Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -5.3826 -0.6380 -2.4177 -0.3884 -2.1183
#>   0.8835  1.0549 -0.2785  0.7745  2.1658
#>  -2.5168 -2.1770 -1.5135 -0.8255 -1.3903
#> 
#> (2,.,.) = 
#>   0.8314  0.6233 -0.1486  0.3952  1.5194
#>   5.8146 -0.3426  3.0218 -0.4587  0.7348
#>   0.3877  0.5245  0.1555 -0.5042 -0.0696
#> 
#> (3,.,.) = 
#>   1.1349  0.2673  0.5653  0.3072  0.5610
#>  -1.4097  2.3617 -0.6879  2.2120  1.9065
#>  -7.6150 -1.4304 -3.7288 -0.5379 -2.6612
#> 
#> (4,.,.) = 
#>  -2.6181 -2.3375 -1.3477 -1.6793 -2.4084
#>   2.7932 -1.4901  2.2547 -1.5051 -2.1813
#>  -1.6897  0.2434 -0.5496 -0.2019 -0.9212
#> 
#> (5,.,.) = 
#>   1.7085  3.5995 -0.6888  2.0199  5.2320
#>  -2.5725 -1.3336 -1.7860 -1.1727 -1.0422
#>  -0.9829 -0.8382 -1.1455 -1.0441 -0.2076
#> 
#> (6,.,.) = 
#>   2.9836  0.7906  2.0218  0.3219  0.3239
#>  -0.6068  1.6881 -0.5214  1.2918  1.6003
#>  -0.8644 -1.1772 -1.5896 -0.7323  0.7250
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]