Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -1.6024 -2.1524 -1.5312 -6.2106 -4.0131
#>   3.0583 -1.1095 -1.3179  0.2029  0.9300
#>   0.4119 -0.1720 -1.0651  2.7917  2.4251
#> 
#> (2,.,.) = 
#>   1.3942 -0.5728 -1.3723 -0.0024 -0.4116
#>  -0.0173  0.3965  1.6780 -1.7756 -1.0828
#>   0.2742 -1.1163 -0.5445 -2.7469 -1.3567
#> 
#> (3,.,.) = 
#>   0.8553  0.4058  1.1118  1.7196  2.2782
#>   1.5218 -0.6045 -0.7009  0.2792  0.7400
#>  -0.3666 -1.2624 -0.5316 -1.8676  0.0919
#> 
#> (4,.,.) = 
#>  -2.3438  0.8953  0.6719  1.0654  0.2400
#>  -1.0509  0.6259  2.0345 -0.6800  0.2027
#>   0.5621  0.6776 -0.5848  2.5881  0.7923
#> 
#> (5,.,.) = 
#>   0.0226 -0.2321 -0.6924  0.5606  0.3973
#>   2.0128 -0.6233 -1.5133  2.0329  1.8745
#>  -2.8150  1.6753  2.3754  1.1616  0.5713
#> 
#> (6,.,.) = 
#>   1.9467 -1.4210 -3.2246  1.1684  0.7262
#>  -0.3352  1.0475  1.7574  1.0552  0.9023
#>   2.0914 -0.8830 -2.0669  1.0574  0.6765
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]