Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   1.2102  0.5104 -1.6399 -0.9342  1.8006
#>  -0.2913  0.0157 -1.7007  1.2344 -1.1248
#>  -1.6630 -1.2470  3.7031  1.1761 -2.6344
#> 
#> (2,.,.) = 
#>  -0.4381  0.7829 -0.9319  0.6183 -1.5752
#>  -1.0047 -3.7066 -0.8615  2.4992  2.2029
#>   0.2858  2.3712  1.3198 -1.0348 -3.1710
#> 
#> (3,.,.) = 
#>  -0.4226  1.9755  0.0677 -0.0356 -3.0482
#>  -1.3363 -1.3351  1.1983  0.7837  0.5430
#>  -1.0713 -2.8598 -4.7626  4.0366  0.7717
#> 
#> (4,.,.) = 
#>  -0.2465 -2.3952 -2.9964  1.5067  3.6285
#>  -0.0597  3.5476 -1.3931 -0.3763 -3.9275
#>  -0.8447 -1.7833 -1.7214  1.2625  2.4890
#> 
#> (5,.,.) = 
#>   0.4356 -0.8693  2.1458 -1.6321  3.0248
#>   0.3361  1.6260  3.8629 -1.1200 -4.1204
#>   0.1977  1.0642  2.8316 -0.9031 -2.5532
#> 
#> (6,.,.) = 
#>  -1.0389 -1.1978 -4.6952  2.4884  1.5052
#>  -0.5300  0.4062 -4.1255  0.9095  1.4974
#>   0.0665 -1.0728 -4.4141  1.4097  2.6031
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]