Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.5553 -1.2923  1.8996 -0.0186  0.1568
#>   2.2200 -1.3636  0.1600 -3.0485  2.0603
#>  -1.2232  1.4655  1.8182  1.6405 -2.2185
#> 
#> (2,.,.) = 
#>   2.9014  3.4237 -0.8667 -2.3948 -0.2281
#>   0.5249  2.5558 -0.3875 -0.5277  1.1579
#>  -2.8086 -2.4131  0.6133  2.5194  0.0063
#> 
#> (3,.,.) = 
#>  -1.1035 -0.3501 -0.2383  0.9064  0.8795
#>  -0.5492 -1.6003 -2.9525  0.6425  1.4497
#>  -2.6353  0.0411  0.9131  2.9599 -1.5375
#> 
#> (4,.,.) = 
#>  -3.2958 -1.3016  4.0351  2.3054  0.2794
#>   1.1235 -0.0329  0.8880 -0.9661 -1.3811
#>  -1.8329  0.1970 -4.0639  2.6572  0.5175
#> 
#> (5,.,.) = 
#>   2.3838  1.3413 -0.7051 -1.4675 -2.4779
#>   0.9596  0.8084  3.0363 -1.3406 -0.8287
#>   1.0531 -3.1663 -5.5444 -0.9813  2.7939
#> 
#> (6,.,.) = 
#>   0.1937 -4.0739 -6.0824 -0.0452  2.7149
#>  -2.3263 -0.4238  3.3967  1.6958 -0.2935
#>  -0.5039 -0.6272  2.9556 -0.0989 -0.3282
#> 
#> (7,.,.) = 
#>   0.3138 -4.3735 -5.0261  0.5763 -0.8359
#>   3.7626  3.2444  0.5606 -4.2877  2.1424
#>   1.4928  3.1702 -1.3334 -1.3447  1.4666
#> 
#> (8,.,.) = 
#>  -2.3062 -0.8531  5.5646  1.2986 -0.7965
#>   0.4744  0.3348 -1.9393 -0.2861  0.9022
#>  -0.8942  0.1253  1.8264  0.2481  1.0094
#> 
#> (9,.,.) = 
#>   2.1000  2.3996 -0.4092 -0.8941 -3.3192
#>  -1.6213  0.7177 -0.2395  2.0944 -1.0436
#>  -0.9339 -2.1097  1.8567 -0.6205  3.3980
#> 
#> (10,.,.) = 
#>   0.7612  1.5738  1.4947 -1.4920  1.8915
#>   3.2780  1.2614 -0.1605 -3.4288  0.7321
#>  -0.8478 -0.5943 -0.7677  1.5266 -1.9130
#> [ CPUFloatType{10,3,5} ]