Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -3.2672  0.0751 -2.8525 -1.4846 -1.1129
#>  -1.3820  0.9775  0.7946 -2.6486  1.8577
#>  -1.1499  1.9378 -0.2180 -2.9622 -1.0472
#> 
#> (2,.,.) = 
#>   1.1085 -1.3416 -0.6998  3.1756 -0.0326
#>  -0.9287 -0.2210 -1.4048  0.8842  0.7029
#>  -0.5341  1.2344 -0.7203 -1.1692 -1.2567
#> 
#> (3,.,.) = 
#>   0.5992  1.2008  3.9878 -4.9429 -0.4104
#>   1.6171 -2.0284  4.8204 -0.9871  3.2017
#>  -3.2997 -0.0968 -2.1696 -1.1746  1.3904
#> 
#> (4,.,.) = 
#>   1.1052  0.9906  1.9653 -1.5213 -0.0264
#>  -1.3162 -2.0862 -1.8298  2.3697  1.4367
#>   0.0021 -0.2480  0.6869 -0.8617 -0.4500
#> 
#> (5,.,.) = 
#>   0.3905 -0.6563 -0.9574  1.6542 -1.0394
#>  -1.5167  1.7698 -0.1636 -2.8717  0.2269
#>   2.8655  1.8522  1.9501 -0.4459 -1.8309
#> 
#> (6,.,.) = 
#>  -1.6171 -1.0606  1.1648 -1.5470  3.2025
#>   0.4204 -0.1249  2.0251 -0.9706  1.9395
#>  -0.5747 -0.3233 -0.2481  0.2123  1.1304
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]