Matmul

Matmul

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   1.4806  0.7357  2.2820  1.7852  0.2051
#>  -1.6206  0.8610 -2.0916 -3.9412 -1.3084
#>   0.3202 -0.8290  0.7629  0.9355  0.7283
#> 
#> (2,.,.) = 
#>   1.0535 -0.0830  0.8026  2.2873  0.4945
#>  -1.1640 -2.3308  1.2600 -0.1002  1.2050
#>   0.0161 -0.8521 -1.2392 -1.0904  0.5303
#> 
#> (3,.,.) = 
#>  -1.9877 -0.6433  0.5851 -2.0378 -0.2922
#>  -0.0874  0.6323 -0.8763 -0.8966 -0.5077
#>   1.8917 -0.0949 -0.1206  2.7672  0.7758
#> 
#> (4,.,.) = 
#>  -1.4333  0.8263  2.2591 -2.3868 -0.9661
#>   1.1743  0.0186  2.1502  2.6324  0.5511
#>   1.4276  0.1860  0.2327  1.8588  0.4327
#> 
#> (5,.,.) = 
#>   1.0135 -0.0394 -1.5802  0.6864  0.3271
#>   0.0136 -2.4832 -0.0113  2.3612  1.6668
#>  -0.5469 -2.4532 -1.7275  0.7369  1.3404
#> 
#> (6,.,.) = 
#>  -0.9275 -1.4724 -3.9998 -1.2705  0.4079
#>   1.4448  0.7117  2.9034  2.9790  0.2267
#>  -0.9437  0.9234 -2.1575 -3.3477 -1.0920
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]