Matmul

Matmul

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.6868 -0.0713 -0.0507 -1.1148 -0.3051
#>  -2.1602 -4.6267  2.0159  2.8009 -1.1730
#>   0.6896  0.0871 -0.0300 -2.9959 -0.3848
#> 
#> (2,.,.) = 
#>  -1.2393  0.1505  0.8043 -1.8427  0.6333
#>  -3.3145 -1.2599  1.9727  2.9336  1.2812
#>  -1.6399 -0.9063  2.1012  3.7124  1.4412
#> 
#> (3,.,.) = 
#>  -0.7927 -1.5939  1.8450 -1.7326  0.0898
#>  -2.9026 -2.8065  3.6334  1.5473  1.1355
#>   0.6109 -0.1144  2.0760  3.5135  1.6053
#> 
#> (4,.,.) = 
#>   0.5053  1.3629 -0.4934  1.1239  0.6414
#>  -3.1645 -2.8169 -0.3725  3.1654 -1.3586
#>  -0.5443 -2.0369  0.0631 -2.3764 -1.5192
#> 
#> (5,.,.) = 
#>  -2.1862 -0.7500  1.2082  5.1230  1.1789
#>   3.0815 -1.2459 -2.5410 -1.2842 -3.1616
#>  -3.5564 -1.2160  1.5073  5.1406  1.2700
#> 
#> (6,.,.) = 
#>   0.7249 -0.1746 -2.6967  1.9995 -1.8311
#>   3.6772 -0.9938 -0.4844 -1.0590 -1.6691
#>   3.0011 -2.9385 -1.4134 -1.5137 -3.5686
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]