Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.9705  3.3883  6.1322  1.3714  0.9559
#>  -0.0747  0.0759 -2.5374  0.7567 -0.9621
#>   0.5326 -0.0473  1.0143 -1.9580  1.5929
#> 
#> (2,.,.) = 
#>  -1.2118  1.5280  1.3588  1.7877 -1.2406
#>   1.5641 -2.2976 -4.5480 -2.2106 -1.2803
#>  -0.9860  1.2615  1.0008  3.2355  2.0360
#> 
#> (3,.,.) = 
#>  -0.4236  0.7049 -0.6003 -0.6963 -2.9963
#>   0.3946 -0.1840 -1.0603  0.3919  2.5644
#>  -0.7177  1.5611  3.1910 -0.1810  0.8663
#> 
#> (4,.,.) = 
#>  -0.4785  0.6107  0.9703  0.7672 -0.0405
#>  -0.5103  1.6959 -1.6632  1.0946  1.5644
#>  -0.4399  0.3477 -1.0626  0.4216 -2.6173
#> 
#> (5,.,.) = 
#>  -0.1917  1.3494 -0.5750 -0.4129  1.3411
#>  -0.4484 -0.9215 -2.4560  4.3978 -0.6747
#>  -1.8219  3.4037  3.1829  1.7593  0.6987
#> 
#> (6,.,.) = 
#>  -1.2612  2.1780  0.3527  1.3080 -1.2689
#>   1.3084 -2.0636 -1.7910 -1.5061  0.5896
#>   0.6812 -1.5598  1.1419 -0.5184  0.8482
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]