Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   2.1323 -1.4490  0.5289 -0.3229 -0.7485
#>   0.1966 -0.8602  2.7374 -1.2696  1.9119
#>  -0.1309  0.3466  0.1101  1.5818 -0.3173
#> 
#> (2,.,.) = 
#>   0.8831 -0.9848 -0.3081 -2.7993  0.1935
#>  -2.5368  1.5602 -0.6568 -0.6096  1.4588
#>  -0.7847  0.7719  1.8274  3.4535  0.7502
#> 
#> (3,.,.) = 
#>   0.0009  0.3335  2.8372  4.6645  0.7148
#>   0.7165  0.2291 -1.0117  2.5973 -1.6985
#>   0.3698 -0.2236 -0.4121 -0.3573 -0.6295
#> 
#> (4,.,.) = 
#>   2.1697 -0.8921 -2.7176 -0.3876 -3.0486
#>   0.3728  0.0908 -0.9097  0.9015 -1.4309
#>  -0.3847  0.1215  1.7971  1.2475  0.8441
#> 
#> (5,.,.) = 
#>  -2.8468  2.2318  2.2867  4.9986  2.0808
#>  -0.9708  0.5705  0.2971  0.1411  0.9667
#>   0.2777  0.1451 -1.0931  0.5479 -0.7227
#> 
#> (6,.,.) = 
#>   0.0197  0.3670 -1.3932  0.6642 -1.0813
#>   0.7916 -0.5653 -0.1717 -0.6590 -0.2915
#>   0.3524 -0.4146  2.0581  0.9342  1.0893
#> 
#> (7,.,.) = 
#>  -0.2483  0.3350  0.9762  1.9122  0.6330
#>  -1.8686  1.4531 -0.8284  0.7999  0.8128
#>  -1.5773  1.5853 -1.5044  1.9342 -0.6694
#> 
#> (8,.,.) = 
#>  -0.9813  0.7444 -0.2745  0.5506  0.1762
#>   3.5680 -2.8675  2.2225 -1.4777 -0.7482
#>   1.4228 -0.9416 -0.4274 -0.7715 -1.2794
#> 
#> (9,.,.) = 
#>   3.1835 -2.8158  1.7351 -2.9524 -0.4997
#>   0.9061 -0.8945 -0.6185 -2.5200 -0.1992
#>  -1.7364  2.4113 -3.1392  4.0749 -1.3058
#> 
#> (10,.,.) = 
#>  -1.1548  1.0403 -1.8754  0.0159 -0.7954
#>  -0.5555  0.7383 -0.0705  2.1449 -0.3996
#>   1.7117 -1.0560 -0.6153 -0.6575 -1.4627
#> [ CPUFloatType{10,3,5} ]