Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.8005 -0.5485 -1.3080  3.2286  1.1779
#>  -2.0621 -1.3834  1.4177 -3.5727 -1.5935
#>   1.7609  1.0368  1.5256 -1.8614 -1.3563
#> 
#> (2,.,.) = 
#>  -0.4201  1.5678 -3.8210  2.2778 -0.2174
#>   0.3371 -1.1676  2.4411 -4.8564  1.9359
#>  -0.5237 -0.9161 -0.4534 -0.5760  2.7657
#> 
#> (3,.,.) = 
#>  -0.3407 -0.3381  1.8035 -4.0629 -1.5238
#>  -0.1323  1.9126 -4.5900  6.0927 -0.5621
#>  -0.7223  1.4622 -3.3651  3.6170 -1.8001
#> 
#> (4,.,.) = 
#>  -0.4577 -0.2378  0.3216 -3.2269  0.1861
#>  -1.2334 -1.0857  1.3528 -3.2912 -0.6125
#>  -1.1661 -0.3527 -1.9731  3.5294  0.7401
#> 
#> (5,.,.) = 
#>   0.2666 -0.6780  1.0134  2.3646  0.3328
#>  -2.0953  1.0252 -3.6505 -1.6970 -1.3730
#>   1.0560 -0.8070  2.8924 -1.9696  0.7192
#> 
#> (6,.,.) = 
#>   0.3308 -1.1540  3.5721 -2.6225 -0.8095
#>  -0.7636 -1.4120  2.5755 -4.1024 -0.3095
#>   0.3175 -0.9797  0.7651  0.2228  2.5637
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]