Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   0.6373 -0.1770 -0.8553 -0.1410  1.0948
#>   0.2653  0.1424 -0.2664 -0.0724  0.3581
#>   1.5223 -0.4284 -1.1659  0.0192 -5.7587
#> 
#> (2,.,.) = 
#>  -1.0702 -2.5600  0.1180  0.1172  0.5414
#>  -0.4568 -1.4047  0.3893  0.1652 -4.1569
#>  -0.2818  1.6077  1.0836  0.1609 -1.7310
#> 
#> (3,.,.) = 
#>   0.9508  1.9410 -0.5061 -0.3187  2.0804
#>   1.0446 -1.1208 -0.9181  0.1751 -5.7069
#>   2.8560  1.6280 -2.6141 -0.4189  1.9719
#> 
#> (4,.,.) = 
#>  -0.6152 -0.9553 -0.1119 -0.1249  3.8294
#>   0.4448  2.8856  0.6144  0.1173  0.0602
#>   1.5041 -0.7689 -1.6834 -0.0810 -1.8125
#> 
#> (5,.,.) = 
#>   0.9559 -0.9798 -0.9112  0.1161 -4.3888
#>  -3.8014 -3.1870  2.3489  0.0924  4.3972
#>   0.2182  3.3156  0.8003 -0.0751  2.0022
#> 
#> (6,.,.) = 
#>  -2.2353 -0.6637  1.9259  0.2580  1.8225
#>  -2.4617 -1.8159  1.7098  0.2832  2.0659
#>  -3.7953 -3.2484  2.5259  0.2976  2.5201
#> 
#> (7,.,.) = 
#>   1.0184  2.1070 -0.4612 -0.1278  1.6924
#>  -0.0118 -0.2290  0.3643  0.2250 -4.1323
#>  -0.4732  2.7039  1.0516 -0.1000  4.3806
#> 
#> (8,.,.) = 
#>   0.8529  3.5499 -0.3561 -0.4599  7.4113
#>   1.9820  0.1570 -1.5900 -0.0911 -4.2541
#>  -2.4012 -0.7808  1.7935  0.1646  4.3320
#> 
#> (9,.,.) = 
#>  -0.5546  1.4242  0.7114 -0.1877  3.8034
#>   0.7484  2.0718  0.4234  0.1299 -4.1603
#>  -0.0512  1.9835  0.0731 -0.3169  6.8517
#> 
#> (10,.,.) = 
#>  -2.2142  1.7744  2.6228 -0.0131  3.5076
#>   1.9915  1.3114 -1.1988 -0.0175 -3.9069
#>  -0.7840  2.6204  1.1148 -0.0530  6.6845
#> [ CPUFloatType{10,3,5} ]