Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.4405 -1.9235  1.8864  0.1482 -1.7042
#>   1.0431  0.4611 -1.6445 -0.1416  1.2879
#>  -0.5010  2.9720 -6.0548  0.0754  3.4671
#> 
#> (2,.,.) = 
#>  -0.9033 -0.2932  1.7017 -0.2551 -0.4472
#>   1.0109 -2.0829  3.8779 -0.9204 -0.4048
#>  -5.1517 -0.5922  3.3554 -1.0803  0.0564
#> 
#> (3,.,.) = 
#>   0.0801  0.7572 -2.7287  0.3353  0.8314
#>   2.0858  1.1585 -1.8261  0.9472 -0.6402
#>  -2.0138  1.8555  0.1738  0.5171 -0.8902
#> 
#> (4,.,.) = 
#>  -2.5720  4.7847 -7.0299  0.1163  4.1163
#>   0.5164  0.1508 -0.5637 -0.4613  1.3056
#>  -0.8124  0.5049  0.1975 -0.9484  1.9169
#> 
#> (5,.,.) = 
#>  -0.4774  1.2248 -1.8833 -0.1134  1.4049
#>   0.5008  1.4073 -2.2900 -0.0432  1.5752
#>   0.6223  0.7434  0.3318 -0.3210  0.7138
#> 
#> (6,.,.) = 
#>  -3.4299 -0.4091  1.2356 -1.5239  2.1927
#>   1.8596  0.7090 -2.6096  0.6276  0.2905
#>  -2.8265  0.1423  1.3150  0.1720 -1.1862
#> 
#> (7,.,.) = 
#>  -1.2115  2.5344 -4.6630 -0.4156  3.6433
#>   0.8884 -0.2996  0.8900 -0.2498  0.0764
#>   0.7651 -1.2457  1.7669  0.2993 -1.7151
#> 
#> (8,.,.) = 
#>  -2.7690 -1.1116  2.2364 -0.8080  0.1086
#>   2.0759 -3.1296  3.9637 -1.0537 -0.3265
#>  -2.0521 -1.5671  1.8339 -0.4565 -0.4681
#> 
#> (9,.,.) = 
#>   2.6670  2.8540 -5.2286  0.3748  2.6505
#>  -2.2068 -2.4354  1.6949 -1.3361  1.2138
#>  -0.5166 -0.0640 -0.1560  0.0421 -0.0632
#> 
#> (10,.,.) = 
#>   4.6366  0.9265 -3.0925  1.3121 -0.6296
#>  -5.5672  2.4594 -0.5836 -0.0290  0.4829
#>   2.1579 -0.3841 -0.9030  0.0181  0.4848
#> [ CPUFloatType{10,3,5} ]