Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.9152  4.9234 -0.5238  1.5758  2.3735
#>  -0.1263 -1.2204  0.3834 -0.2303 -1.0474
#>  -2.2036  0.0090 -0.2840  0.5337  1.5938
#> 
#> (2,.,.) = 
#>   3.0608  1.4155 -3.1896 -0.7340  1.8736
#>   0.2944  1.5400  0.4512  0.4827 -0.9084
#>   2.0962 -0.6564 -1.1091 -0.7714 -0.9713
#> 
#> (3,.,.) = 
#>   1.3822 -0.7413  0.5507 -0.5417 -1.2009
#>   0.8432  0.3942 -1.5152 -0.2452  0.9199
#>   1.2174 -0.6733  2.0149 -0.2648 -3.1099
#> 
#> (4,.,.) = 
#>  -1.1638  5.5507 -1.4756  1.7604  3.2018
#>  -1.7390 -0.6872  2.9606  0.5448 -1.9024
#>  -1.0072  2.0144  1.1871  0.9239  0.2621
#> 
#> (5,.,.) = 
#>   0.1038  1.3620 -0.6731  0.3071  0.7902
#>   0.8741  0.6244  0.4198  0.0030 -0.8036
#>  -0.2153 -3.5609 -0.8096 -1.2019  1.5156
#> 
#> (6,.,.) = 
#>  -1.7259 -1.2109  1.1381  0.1942 -0.2032
#>  -0.5916  1.3818  0.6723  0.6352 -0.1418
#>  -1.4332  1.5505  1.4188  0.8921  0.4832
#> 
#> (7,.,.) = 
#>  -1.8884 -7.2552  1.6884 -1.3927 -2.5496
#>   0.3347  0.0645 -1.1678 -0.1372  0.4520
#>   0.9962 -4.4989 -1.6153 -1.7085 -0.1456
#> 
#> (8,.,.) = 
#>  -3.8796  1.2850  1.7624  1.5307  1.0046
#>   1.0739  3.1282  1.2709  0.6962 -0.5412
#>   2.2956  0.7955 -3.1227 -0.7142  2.1569
#> 
#> (9,.,.) = 
#>  -1.0517  0.2822 -0.0725  0.4067  0.1322
#>   1.2284 -2.3971  1.0849 -0.8875 -2.3537
#>   1.2838  3.4223  0.2981  0.6656 -0.0392
#> 
#> (10,.,.) = 
#>  -0.0429  1.4059 -1.7565  0.2885  1.5087
#>  -1.0410  0.3352  1.1773  0.5018 -0.6113
#>   0.0173  2.0730  0.2994  0.7323 -0.8734
#> [ CPUFloatType{10,3,5} ]