Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -1.3763  1.0948 -0.3655  2.3697 -1.9411
#>   0.6945 -0.4453  0.0497  0.6040  3.6512
#>  -1.9802 -2.5779 -2.0313 -2.2877  0.3562
#> 
#> (2,.,.) = 
#>   0.2253 -1.6133 -0.0742 -2.8821 -0.2793
#>   1.6935 -1.0909  0.4883  0.0235  5.9736
#>  -0.7595  0.0128 -1.2970  1.5183  2.9684
#> 
#> (3,.,.) = 
#>   1.8959 -0.9657  0.8258 -3.3951  0.7230
#>  -0.3079  0.8550 -0.5410 -0.2056 -1.4517
#>  -2.4317 -2.2794  0.1412  0.0526 -3.5666
#> 
#> (4,.,.) = 
#>   0.6212  2.1788 -0.2539  2.0483  1.4649
#>  -0.0834 -1.6088 -0.1774 -1.8063  0.5393
#>   0.8256  1.1567 -0.6049  1.1067  3.4644
#> 
#> (5,.,.) = 
#>  -1.0567 -1.5100 -1.4256 -1.6692  0.6352
#>   2.5566  0.7429  1.8372 -0.4416  1.3432
#>  -1.7334 -0.9690 -0.9667 -1.2972 -3.0012
#> 
#> (6,.,.) = 
#>  -2.6097 -0.8944 -2.2146  0.7092  0.0672
#>   0.3878 -1.6611 -0.8070 -3.3995  1.3627
#>   0.0813  0.5345  0.1267  1.3865  0.9268
#> 
#> (7,.,.) = 
#>   2.0950  0.4004  1.2686 -1.0549  1.0787
#>   0.5570  0.0336  1.7072  1.3904 -0.5723
#>   0.7365 -0.8971 -0.3034 -2.5756  0.9767
#> 
#> (8,.,.) = 
#>   2.2182  0.4298  2.2620 -0.1709  0.2212
#>  -0.5611 -0.0510 -0.8610 -0.6482 -0.4293
#>   1.0825  1.1848 -0.1623  0.6668  2.4962
#> 
#> (9,.,.) = 
#>  -2.0179 -1.0033 -0.6970 -0.7006 -3.6246
#>  -0.4455  0.3045 -0.7135  1.5745  2.0530
#>   1.3313  0.0099  0.4672 -1.0884  1.3658
#> 
#> (10,.,.) = 
#>   1.8668 -1.1768  1.1478 -2.8461  1.0027
#>   0.6458 -0.0748  0.6731 -0.0447  0.3970
#>  -1.0310 -0.7054 -1.0628 -0.1798  0.4731
#> [ CPUFloatType{10,3,5} ]