Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.7095 -0.1247  2.1053 -1.9175  0.3274
#>  -1.4227  0.2545  2.1167 -1.3235  0.5101
#>   1.0012 -0.1888 -1.2749  1.4202  0.1791
#> 
#> (2,.,.) = 
#>   5.7735 -0.9299  1.7260 -1.0779  2.4923
#>   0.0204  0.1116 -1.6038  0.9302 -0.7719
#>  -2.6884  0.3603 -1.5657  1.2449 -1.2999
#> 
#> (3,.,.) = 
#>   0.0584  0.6572  1.8088 -0.5727  1.1266
#>  -1.9972 -0.3782  1.1881 -1.7309 -0.6772
#>   0.1367  1.0485  1.0998  0.7400  1.3981
#> 
#> (4,.,.) = 
#>   1.8793 -1.4757  0.1053 -1.1238  0.0147
#>   3.0084 -2.2261  1.8689 -4.7262 -0.5615
#>   0.3095  1.0298 -1.1598  1.4771 -0.0324
#> 
#> (5,.,.) = 
#>   3.1825  0.1277  1.3035 -1.0534  1.2754
#>  -0.7681 -0.1298  0.7038  0.4177  0.7241
#>   1.9442  0.4756 -0.8151  1.2272  0.6234
#> 
#> (6,.,.) = 
#>   1.6762 -0.8250  2.4233 -2.1117  1.2269
#>   3.3777 -1.2878 -0.9544 -1.6518 -0.9082
#>   0.1226 -0.2626  0.7610 -0.3649  0.4850
#> 
#> (7,.,.) = 
#>   1.5683 -1.2274  3.8077 -5.0160  0.4197
#>  -0.0018  0.3009 -0.1069  1.6269  1.0177
#>  -4.3841  1.5608  1.8052  0.5495  0.5658
#> 
#> (8,.,.) = 
#>  -4.9112  1.3004 -0.7658  2.9424 -0.1295
#>   1.8738 -0.0312 -0.2734  0.6602  0.7496
#>  -0.2917 -1.2295  1.2877 -1.9680 -0.1705
#> 
#> (9,.,.) = 
#>   2.7989  0.4751  0.9406  1.9089  2.9003
#>   0.6568  0.4332 -0.6366  2.0741  1.0301
#>  -0.5544  1.2279  1.4295  1.6699  2.1337
#> 
#> (10,.,.) = 
#>  -5.4392  0.2318 -2.7612  1.9060 -2.6377
#>   9.2209 -1.4047 -2.2002 -0.5282  0.4207
#>   2.8481 -0.1734  0.3323  0.0429  1.1385
#> [ CPUFloatType{10,3,5} ]