Matmul

Matmul

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   1.3048  1.2849  0.9104  0.6749  0.6009
#>  -1.2395  1.7542 -1.5638 -0.3625  1.1067
#>   2.7924 -0.5955 -1.7831  1.9304 -1.7293
#> 
#> (2,.,.) = 
#>   2.0694  0.1344 -2.2686  0.5777 -0.0222
#>   4.1705  1.4308  1.0111  1.6451  0.5473
#>   5.5884 -1.6424 -0.2293  2.1343 -1.7001
#> 
#> (3,.,.) = 
#>   3.6937  0.4846  1.8319  1.5977 -0.1445
#>   3.6942 -1.4704 -0.4997  1.3468 -1.3359
#>   0.0795  0.8971 -1.3711 -0.2335  0.8069
#> 
#> (4,.,.) = 
#>  -3.7996  0.1582 -1.6903 -1.5678  0.5239
#>  -0.0846 -2.0461  0.5008  0.0841 -1.4776
#>   0.5231 -0.6078 -0.1035  0.2600 -0.5375
#> 
#> (5,.,.) = 
#>  -0.6887  1.0616 -1.4630 -0.4665  0.9264
#>   0.0573 -2.4232 -0.8460 -0.3495 -1.2984
#>   1.8169  0.2445  1.6734  0.9436 -0.1927
#> 
#> (6,.,.) = 
#>   0.9068  1.3132  2.7279  0.0219  1.3091
#>  -0.8029  2.9202  0.6100 -0.0384  1.8100
#>  -3.7792  0.4399  1.4897 -1.8332  1.1941
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]