Matmul

Matmul

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.9270 -0.3964  0.3969 -1.2430 -1.1749
#>  -0.3059  1.8893  1.9639  0.1744  1.0980
#>   0.4167  0.2011  2.7205  3.6700  1.3064
#> 
#> (2,.,.) = 
#>  -1.9615  3.7084  1.3365  0.6976  1.5818
#>  -1.4367  0.5814 -0.9268  2.1059  0.8056
#>   5.5841 -4.5737  3.6511 -5.1079  1.3731
#> 
#> (3,.,.) = 
#>  -0.6212  2.7019  2.0282 -0.0317  0.5256
#>   0.1751 -0.7950 -0.6344 -1.7119  1.6134
#>   2.4704 -1.9622  1.0978 -3.1294 -0.2142
#> 
#> (4,.,.) = 
#>  -2.5534 -1.4941 -4.3927  4.7447  0.4967
#>  -0.2328  2.8224  3.2965  0.0743  1.7145
#>  -0.2058 -1.6939 -1.3344  2.7620 -0.2375
#> 
#> (5,.,.) = 
#>  -0.1789 -0.0537  1.0971  2.7037  1.5364
#>   0.8456 -2.0639 -1.2415  2.4763 -4.1036
#>   1.2460 -1.6078  0.2521 -2.2071  1.7043
#> 
#> (6,.,.) = 
#>   1.5420 -0.1310  2.6008  0.6805 -0.8307
#>   0.1981 -1.3874 -1.3761 -0.1327 -0.4604
#>  -0.6493 -0.7574 -2.0957  0.0948 -0.3670
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]