Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.8012 -0.6592 -2.8347  4.0414  4.2732
#>  -1.0795 -0.3805 -0.2920 -0.8369  0.2862
#>   0.8521 -3.0431 -0.0216 -3.4999 -0.5172
#> 
#> (2,.,.) = 
#>   1.9768 -5.6740 -0.3852  1.3092  1.8359
#>  -0.4641  1.8588  2.5765  0.0657 -2.7690
#>  -2.4802  0.1427  0.2293  3.2042  1.2294
#> 
#> (3,.,.) = 
#>  -2.7995  0.9728 -2.0416 -2.4332  1.3771
#>  -0.9058 -0.5543  0.2056 -1.2273 -0.2947
#>   0.2421 -1.7446  0.0881  0.7497  0.5484
#> 
#> (4,.,.) = 
#>  -2.4207 -0.2205 -0.8867 -1.0758  0.9634
#>   2.3920  2.3820 -1.7256  0.7326  0.8448
#>   1.3847 -3.6115  1.6713  0.3419 -0.8090
#> 
#> (5,.,.) = 
#>  -2.9815  0.1579  0.2481  0.7285  0.4922
#>   1.6347 -0.4419 -0.8323  2.6008  1.4449
#>  -3.9288  2.8015 -0.9061 -3.2478 -0.2028
#> 
#> (6,.,.) = 
#>   0.9317  2.9011  0.0923  2.0911 -0.2870
#>   0.1218 -2.6268  0.6899  0.3753  0.1017
#>   2.5037  2.9380 -2.9652  5.4885  3.3946
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]