Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.9493 -0.1869 -0.1631  3.1089 -0.9173
#>  -5.7749 -1.2723  1.3829 -0.3706 -2.9075
#>   0.8797  1.0493 -1.6627  2.6533  0.5063
#> 
#> (2,.,.) = 
#>  -3.4988 -1.4284  1.0680  1.1427 -1.7421
#>   0.9465  0.5235 -0.0824 -0.6242  0.1991
#>  -1.3205  1.3191 -2.0611  1.1493  0.0388
#> 
#> (3,.,.) = 
#>  -0.9871 -2.4877  3.2594  0.4895 -1.7407
#>  -1.0309 -1.0163  1.7257 -1.5785 -0.9631
#>   6.3417  2.3777 -2.0077 -0.0239  2.8953
#> 
#> (4,.,.) = 
#>   0.0215 -0.2367  0.5706 -0.4700 -0.2566
#>  -0.1053  1.5500 -1.8273  1.8622 -0.0518
#>  -4.2176  0.6525 -1.0608 -0.6595 -1.2943
#> 
#> (5,.,.) = 
#>  -0.9964  1.5799 -2.9495  2.9362  0.4305
#>   0.8696  1.2460 -1.6932  2.0677  0.5010
#>   5.7824  0.7327 -0.1507 -1.7829  2.6648
#> 
#> (6,.,.) = 
#>  -0.4618 -1.3268  1.8640  2.9047 -1.7358
#>   2.7329 -0.2738  1.2636 -0.9271  0.4300
#>  -0.1019  0.9583 -0.0354 -1.6106 -0.5032
#> 
#> (7,.,.) = 
#>  -1.2142 -0.3720  0.5633 -0.7003 -0.6413
#>  -0.9740 -0.2954 -0.0587  1.2809 -0.4527
#>  -3.0806 -2.4834  3.7497 -3.5557 -2.1919
#> 
#> (8,.,.) = 
#>  -1.5334  0.0033 -0.4620  1.4343 -0.6073
#>  -0.4793 -0.2650  0.3510  1.3489 -0.7106
#>   2.6627  1.6438 -2.7589 -0.1384  2.7075
#> 
#> (9,.,.) = 
#>   1.1516  2.2371 -3.7259  2.7312  1.7102
#>   1.5656  0.4951 -0.4822 -1.6176  1.2164
#>  -0.4522 -3.2757  4.7282 -0.9825 -1.9331
#> 
#> (10,.,.) = 
#>   0.9731  2.0939 -2.5044  0.6440  0.9928
#>   1.3163  1.2658 -1.5567 -1.0086  1.3616
#>  -3.1844  0.3813 -0.4186 -0.2290 -1.3563
#> [ CPUFloatType{10,3,5} ]