Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -2.0271 -0.2149 -1.2504  3.9989 -0.0802
#>   0.3190  0.7490 -0.1358 -1.9446 -0.6521
#>  -3.3108  1.3626 -2.8875  2.5360 -1.8795
#> 
#> (2,.,.) = 
#>   0.6201 -1.3175  1.7261 -0.8218  0.5984
#>   2.2821  0.5118  1.5446 -5.4812 -0.3612
#>  -0.2414 -1.9516  2.2920  2.8496  0.9449
#> 
#> (3,.,.) = 
#>   1.1220 -0.2664  0.6737 -1.5554  0.4681
#>   1.4242  4.0481 -3.3609 -5.3369 -1.6740
#>  -0.0081  0.8187 -0.7010 -1.2736 -0.5960
#> 
#> (4,.,.) = 
#>   0.8125  0.7127 -1.5754  0.2994  0.7674
#>   0.0241  0.9930 -0.5457 -0.5774 -0.6143
#>  -0.6889  1.4725 -3.4230 -0.2011 -0.3222
#> 
#> (5,.,.) = 
#>   1.2930  1.6542 -1.3498 -3.9091 -0.5087
#>   0.0798 -0.9176  2.1683  0.7020  0.0403
#>   1.9470  1.3370 -0.8122 -2.9661  0.2085
#> 
#> (6,.,.) = 
#>  -0.8085 -2.3857  2.8500  5.6781  1.1990
#>  -0.9125 -0.1630 -0.1490  1.8634 -0.1339
#>  -0.6582  1.3443 -3.2837  0.4293 -0.1289
#> 
#> (7,.,.) = 
#>   1.7552 -0.5171  2.2508 -4.5786 -0.0982
#>  -3.3325  0.6284 -2.7909  5.0056 -0.8882
#>  -0.1452 -0.7941  2.4504  0.4801 -0.3753
#> 
#> (8,.,.) = 
#>  -2.5582 -0.3313 -1.5335  4.6340 -0.1597
#>   2.9672 -1.1270  4.2703 -3.4697  0.7215
#>  -0.5372  0.7684 -1.3629  0.0805 -0.4280
#> 
#> (9,.,.) = 
#>  -1.6370  0.5824 -1.4596  2.6769 -0.5566
#>   2.3690 -1.3776  3.3677 -3.1114  0.8697
#>   0.7242 -1.8432  2.2567  2.2833  1.4978
#> 
#> (10,.,.) = 
#>   0.4091  0.0563  0.2539  0.0895  0.1914
#>   0.9589  1.8700 -2.2453 -5.5285 -0.9104
#>  -1.5421 -0.5838 -0.3939  2.5293 -0.0751
#> [ CPUFloatType{10,3,5} ]