Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   1.2107 -1.9933  0.7064 -1.2436  1.1962
#>   0.2281  0.3725 -0.5968  1.4756 -0.9518
#>  -0.8770 -3.8745  0.6301 -1.7164 -0.7970
#> 
#> (2,.,.) = 
#>   0.6345 -3.6043  0.6638 -1.3351  0.1807
#>  -1.9993  0.9825 -0.3396 -0.0906 -0.8834
#>  -1.7139  5.1907 -0.4566  1.7233 -1.7149
#> 
#> (3,.,.) = 
#>  -3.8278  3.0602 -1.3770  0.6358 -1.1844
#>   0.1675  1.5291  0.0458  0.0572  0.6080
#>  -0.8358 -6.0343  0.4095 -1.8316 -1.0100
#> 
#> (4,.,.) = 
#>  -0.9867  1.8693 -0.6497 -0.0826  1.1107
#>   0.8244 -3.0697  0.7470 -0.6267 -0.7570
#>  -1.8198 -0.6801 -0.4966  0.3534 -1.8143
#> 
#> (5,.,.) = 
#>  -0.9320 -0.1237 -0.1582 -0.7247  0.5016
#>   1.3752 -4.4394  1.2690 -1.4469 -0.3562
#>  -3.1924  4.9318 -1.3515  0.4987  0.4153
#> 
#> (6,.,.) = 
#>  -0.9969 -3.0731  0.3249 -0.7480 -1.6434
#>   0.6487 -1.1491  0.0053  0.0187  0.1703
#>   1.7757 -0.8563  0.6406  0.4489 -0.4785
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]