Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   1.9012 -0.8697  2.1002 -0.9527 -0.0133
#>  -1.3985 -0.5399 -1.4121 -0.8773  2.2093
#>   1.4401  1.1525  1.7637 -0.5784 -1.4876
#> 
#> (2,.,.) = 
#>  -0.0678  6.1007  1.6276 -2.4436 -0.1148
#>   0.1755 -0.5024  0.1004  0.3415  0.0444
#>  -0.8268 -2.2531 -1.5257  0.2463  0.7936
#> 
#> (3,.,.) = 
#>   0.0537  3.2650  0.7577  1.2269 -1.4622
#>  -1.9649 -0.8351 -2.1669  0.0113  2.2407
#>  -0.4707  1.2595 -0.2169 -0.7161  0.1351
#> 
#> (4,.,.) = 
#>  -2.5914 -3.5647 -3.2994 -0.5236  4.2670
#>  -3.2250 -1.2026 -3.7865  0.5689  2.5053
#>  -0.6762 -0.2881 -0.5949 -1.4593  1.6520
#> 
#> (5,.,.) = 
#>   1.2409  2.6557  1.6464  0.0286 -2.8427
#>  -1.2840  0.3158 -1.6917  1.3734 -0.8564
#>   1.2298 -3.3950  0.8072 -0.0346  1.0780
#> 
#> (6,.,.) = 
#>  -1.2408  2.1462 -1.1212  0.5346 -0.8644
#>  -1.9146  0.3192 -1.5323 -1.4112  3.3454
#>   0.3405 -0.3353 -0.0242  0.6013 -1.4575
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]