Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -0.3824  0.0454  1.0631 -0.5170 -2.9853
#>  -0.4579 -2.8684 -1.6437  0.7040 -0.7168
#>  -1.1962  1.8101  0.0400 -1.2815 -1.1514
#> 
#> (2,.,.) = 
#>  -1.0669  2.1964  0.5232 -1.8929 -2.3436
#>  -1.8082  0.5365 -2.4142 -1.8001 -0.7280
#>   1.1673  1.5785  2.8188  0.5166  0.0455
#> 
#> (3,.,.) = 
#>  -0.1168 -2.9976 -3.6923  1.2100  3.6381
#>  -1.8275  0.8911  0.9940 -1.8295 -5.6746
#>  -0.8866 -0.6958 -1.1093 -0.4751 -1.0956
#> 
#> (4,.,.) = 
#>  -1.3705  2.8340 -0.3511 -2.0481 -0.6725
#>   1.1550  1.0181 -0.8418  0.9759  5.8523
#>   0.3327  0.3630  0.6145  0.0715  0.0477
#> 
#> (5,.,.) = 
#>   3.0690  0.8474  5.0826  2.3521  0.9009
#>   2.7504 -1.5241  4.4816  3.0929 -0.5299
#>  -0.3576  0.3315 -0.2373 -0.7170 -0.8646
#> 
#> (6,.,.) = 
#>   2.5340 -1.3893  1.9259  2.8653  3.0388
#>   2.3752 -1.3948  3.1566  2.4332  0.1188
#>  -3.8725  1.0416 -4.2082 -3.6393 -2.8334
#> 
#> (7,.,.) = 
#>   1.6082  3.3781  1.8273  0.2667  4.2094
#>   1.2110  0.8410  0.3824  0.5038  2.8607
#>  -0.2111  0.9323  0.7072 -0.6019 -1.1231
#> 
#> (8,.,.) = 
#>  -1.1494 -2.1314 -0.2835 -0.2561 -4.3381
#>  -1.7707  2.3644  0.4703 -2.3231 -3.5872
#>  -3.7149  0.1787 -4.4333 -3.2519 -2.8258
#> 
#> (9,.,.) = 
#>    2.2613  -1.7536   1.5211   2.5348   2.2482
#>    4.7889   1.7949   4.4730   3.5003   7.2254
#>    0.2683   3.4682  -3.8875  -0.4801  10.5057
#> 
#> (10,.,.) = 
#>   0.6379 -0.5063  2.4116  0.7668 -2.4159
#>   0.5451  0.8160 -1.2754  0.3746  4.4692
#>  -0.8787  1.1648 -0.4451 -1.3223 -1.0334
#> [ CPUFloatType{10,3,5} ]