Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   2.6912  0.6684  0.0077 -2.0448  1.3169
#>   3.0235  0.4318  1.0560 -2.6110  1.5164
#>  -2.0322 -0.3606 -0.6679  0.4272 -0.3935
#> 
#> (2,.,.) = 
#>  -0.3130 -0.3616  0.4217  0.7107 -0.3648
#>   1.0493  2.1165  2.1972 -4.0881  0.9962
#>   2.1983  0.8679  1.7151 -2.8454  1.1718
#> 
#> (3,.,.) = 
#>   6.9368 -0.5268  0.5034 -3.4526  3.1798
#>  -5.9845 -0.4038  2.2661  6.5228 -4.6420
#>   1.1418  0.2946  0.7378 -1.6138  0.7518
#> 
#> (4,.,.) = 
#>   4.8751  1.0307 -1.0880 -6.0252  3.7330
#>  -2.5444 -1.8445 -1.2880  3.8086 -1.4646
#>   0.1909 -1.1010 -2.2726  0.1308  0.7889
#> 
#> (5,.,.) = 
#>  -0.4622  1.3906  2.7599 -1.3782 -0.4641
#>  -3.1566 -0.2803 -0.7261  1.6426 -1.2140
#>  -4.5379  0.5595  0.9015  2.5430 -2.5263
#> 
#> (6,.,.) = 
#>  -1.9487 -0.4331  0.4687  1.9746 -1.2924
#>  -1.4128  0.6380  1.8604  0.5835 -1.1553
#>   0.3288  0.4368 -2.4027 -2.8899  1.7447
#> 
#> (7,.,.) = 
#>  -1.6207  0.4147 -0.2447 -1.5368  0.2687
#>   3.6346 -1.3972 -2.7909 -0.6623  2.1082
#>  -4.2278 -1.4733 -0.9048  4.2753 -2.2443
#> 
#> (8,.,.) = 
#>  -4.6659  1.0276  0.5889  1.3357 -2.0854
#>  -7.9012 -1.1149  0.6141  7.2158 -4.8234
#>  -5.0120  0.5646 -0.1734  2.0960 -2.2108
#> 
#> (9,.,.) = 
#>  -3.5061  0.4300 -0.2430  0.2192 -0.9590
#>   1.2511 -2.5324  0.5027  5.0264 -1.3314
#>  -2.2676  1.0210 -1.0871 -1.2757 -0.0187
#> 
#> (10,.,.) = 
#>   4.8198 -0.0685 -2.4845 -4.1395  3.4835
#>   0.3867  2.2853  2.5537 -4.0400  0.7034
#>   5.2475  0.7234  0.5285 -3.1570  2.2681
#> [ CPUFloatType{10,3,5} ]