Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.9659 -1.1176  0.0228 -0.9884  0.5842
#>  -2.9711  2.1108  0.5161  1.6051 -1.7054
#>   0.7876  3.1140 -2.8025  1.6680  2.0646
#> 
#> (2,.,.) = 
#>  -2.0299  3.4599 -2.9153 -1.3541  0.8139
#>  -0.7507 -0.7176  0.2619 -1.8907 -0.7161
#>   1.0462 -0.6176 -0.4596 -0.7999  1.1766
#> 
#> (3,.,.) = 
#>  -3.3387  3.0724 -1.0459  0.0060 -0.7220
#>  -1.2934 -0.2732  1.0264 -0.3098 -2.4734
#>  -1.0342  0.8675 -0.4478 -0.5376 -0.8987
#> 
#> (4,.,.) = 
#>   0.9431  1.4665 -1.0379  2.0409  0.9627
#>  -2.4148  1.7797  0.3140  1.4055 -0.2060
#>  -1.0898  1.7168 -1.0808 -0.0509  0.1030
#> 
#> (5,.,.) = 
#>  -0.9751 -0.9894  2.4815  1.5669 -3.0658
#>   0.1805 -1.0048  1.7750  1.7417 -0.9744
#>   0.3288  0.7614 -1.1883 -0.4160  0.9434
#> 
#> (6,.,.) = 
#>   2.0141 -0.2107 -2.1560 -2.3514  1.5100
#>  -1.8834  3.4376 -2.3373 -0.1743 -0.1476
#>  -2.7199  4.4204 -2.1514  0.8772 -2.2829
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]