Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -3.8562 -2.0520  2.4338  0.0919  1.1781
#>  -1.8916 -1.5735  2.3732  5.0339  3.3493
#>   1.4687  1.4914  1.2151  6.1360  4.7562
#> 
#> (2,.,.) = 
#>  -0.6255 -0.2516  0.7825  0.9543  0.7371
#>  -2.2275 -2.0156 -0.1109 -2.9679 -1.8085
#>  -0.6180 -0.7360  0.4476  0.6324 -0.1047
#> 
#> (3,.,.) = 
#>  -0.1430  0.4763  0.6588 -0.1949 -0.4205
#>  -1.0005  0.0710  0.5119 -1.9001 -0.7024
#>   1.0196  0.6461 -0.6465  0.1479  0.2333
#> 
#> (4,.,.) = 
#>   0.2743  0.3430 -0.1461  0.4991  1.1315
#>   1.2743  1.5062 -0.3354 -0.1167  0.3281
#>   0.7657  2.7302  2.6782  5.0688  4.6674
#> 
#> (5,.,.) = 
#>  -0.4259 -1.7759 -0.5099  0.9731 -0.2683
#>   0.5071  1.1709  0.4097  0.5521  1.0065
#>  -1.2189 -0.7522  0.3979 -0.9574 -0.3511
#> 
#> (6,.,.) = 
#>   0.0802 -1.2447 -1.2516  0.4141  0.6335
#>   2.4173  1.7541 -0.2019  3.3865  1.9902
#>  -0.5203  0.3290  1.0664  0.9476  1.0856
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]