Matmul

Matmul

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.6827  2.2775 -0.9836  0.1100 -1.5570
#>   1.7818  0.2269 -0.4110  2.6043  2.0581
#>  -1.3928 -0.3561 -3.1273 -1.8589 -3.2131
#> 
#> (2,.,.) = 
#>  -2.9156 -0.8392 -3.6944 -2.5918 -1.7108
#>   0.5513  0.5741 -1.7936  1.7827  1.3070
#>   0.7945  0.5529  1.1628  1.9506  3.1445
#> 
#> (3,.,.) = 
#>   1.3959  3.0146  2.8860  3.3801  3.6657
#>  -2.2228  1.2755  0.4475 -1.9746 -1.3453
#>   1.7808 -1.7063  1.3244  0.4681  0.2973
#> 
#> (4,.,.) = 
#>   1.1948 -0.7655  1.9926  0.0395 -0.1636
#>  -1.1416 -0.4039  2.0284 -1.9370 -0.5257
#>  -0.7624  0.1634 -3.4233 -0.2071 -1.3480
#> 
#> (5,.,.) = 
#>   3.4566 -2.1988  1.6490  3.4939  4.8456
#>   0.0468 -3.2832  1.2398 -2.4749 -1.4631
#>   0.8963 -2.5342  0.3947 -0.4517  0.2286
#> 
#> (6,.,.) = 
#>  -3.1242  0.5574 -2.5013 -3.3665 -3.9203
#>  -0.2161  1.4796  0.5748  0.2315 -0.3076
#>   0.3793 -2.1138 -0.1181 -1.3880 -1.7483
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]