Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   1.2666 -0.5709 -0.7704  0.4736 -2.2975
#>  -2.1579  0.0480  2.9487  1.8239 -3.7011
#>  -1.3630 -1.4584  3.2756 -2.4503 -0.4543
#> 
#> (2,.,.) = 
#>   0.1858  1.0179 -1.3495 -0.9349  3.7501
#>   1.2653  0.5069 -1.4008  0.9927 -2.0393
#>  -3.2198  0.6012  3.0636  0.3459  1.1972
#> 
#> (3,.,.) = 
#>  -0.7195  0.3774 -0.2548  3.3605 -1.5212
#>   0.0526  0.8747 -0.7349  1.2270 -0.3883
#>   1.8771 -2.2067  0.9280 -5.8927  0.4013
#> 
#> (4,.,.) = 
#>  -0.3234  1.0180 -0.6655  2.6460 -1.2921
#>  -3.9838  1.1472  4.7527 -1.3452  0.6172
#>   0.5946  0.7183 -0.0214 -1.9261 -0.0858
#> 
#> (5,.,.) = 
#>  -0.6107 -0.6135  2.3978 -4.1344  1.0358
#>   1.6411 -0.0020 -1.1118 -2.0064  0.2241
#>  -0.8069  0.5553  2.0590 -1.3708 -1.9737
#> 
#> (6,.,.) = 
#>  -3.5422  1.9166  0.6580  5.6400  1.3149
#>   0.1794  0.2393 -0.7554  0.8164  0.3297
#>  -0.0556  0.6358 -1.0807  2.3865 -0.3122
#> 
#> (7,.,.) = 
#>  -2.5933  1.3675  2.5498 -0.5399  1.0415
#>  -2.5624 -1.4876  4.6200 -4.3291  2.2895
#>  -0.8790  0.1007  0.2290  2.2298 -0.6537
#> 
#> (8,.,.) = 
#>   1.5524  0.0886 -3.5427  4.4486 -0.8909
#>  -0.8810  0.2187  0.9088  1.0722 -1.2389
#>   0.7418  0.2543 -1.4286  1.0175  0.0796
#> 
#> (9,.,.) = 
#>   0.9819  0.0341 -1.6897  0.1339  1.3889
#>   0.9504 -0.4638 -0.9801 -0.4302  0.5628
#>  -0.6244  0.6541  0.5083 -1.6394  2.3025
#> 
#> (10,.,.) = 
#>   2.2238 -0.2179 -1.9879 -1.7621  0.6544
#>   2.4810 -1.3667 -0.1522 -1.6080 -4.9583
#>  -0.5136  1.8469 -1.3050  1.1053  2.6329
#> [ CPUFloatType{10,3,5} ]