Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   2.7380 -0.7968  0.8734 -0.8695  4.0417
#>  -1.9915  3.0731 -1.0124  0.7094  1.5933
#>  -3.4209 -0.3444 -3.2848  0.0114  5.7183
#> 
#> (2,.,.) = 
#>   3.1477  2.5955  5.2900  1.1750 -7.2526
#>   4.8215 -3.3936 -0.1198 -2.4166 -0.7097
#>  -5.8123  3.7596  1.3676  3.4061 -4.1394
#> 
#> (3,.,.) = 
#>  -3.1789  4.4921  0.1540  1.8362  6.6485
#>   5.1710 -2.8734  2.9838 -1.1869 -9.1511
#>  -5.6497  1.9701 -1.1470  2.0924 -0.3419
#> 
#> (4,.,.) = 
#>   0.8895  0.1568  1.8785  0.4408 -3.8416
#>  -1.5934  3.3493 -0.9054  0.6040  8.4295
#>  -1.0259  1.5863 -2.5089 -0.4832  5.8661
#> 
#> (5,.,.) = 
#>  -2.4353  6.7011 -1.4609  1.0534  6.6455
#>   3.4415 -0.5893 -1.3463 -2.0814  0.1947
#>   1.7667 -7.3954  0.3510 -1.3103 -5.7176
#> 
#> (6,.,.) = 
#>   1.1759  2.0958 -2.3940 -1.3128  4.5328
#>   5.0558 -4.5348  3.5251 -1.0765 -5.1840
#>   4.2057 -3.4941  1.9737 -1.2805  0.0551
#> 
#> (7,.,.) = 
#>  -1.4203  3.3575 -4.7170 -1.0764  3.7223
#>  -2.8072 -5.1693  0.8738  1.0567 -3.5819
#>   2.5442  0.0532 -1.6838 -1.7913  6.2918
#> 
#> (8,.,.) = 
#>   2.9461 -2.1039  1.8832 -0.6459 -3.9286
#>  -1.3479  2.1834 -4.4459 -1.1113  7.0873
#>   3.8591  2.0065 -2.1980 -2.3708  4.0328
#> 
#> (9,.,.) = 
#>   0.8368  1.5260  0.0100 -0.2026  1.5733
#>   3.7872 -0.4363  1.4656 -1.0161 -3.4041
#>   0.4852  2.4473 -2.2229 -0.9116  2.5610
#> 
#> (10,.,.) = 
#>  -3.7966  4.8138  2.4730  3.1242 -2.8554
#>  -2.2045 -2.9226 -0.2364  0.5481 -1.5892
#>   2.7543 -2.3612  1.7098 -0.6695  0.4427
#> [ CPUFloatType{10,3,5} ]