Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   1.4947 -1.3871 -1.0989  0.7247 -1.3257
#>  -4.4834  4.1667 -2.2363 -0.8793 -2.1580
#>  -8.0405  6.2060 -1.3015 -1.7159 -1.0962
#> 
#> (2,.,.) = 
#>  -7.7862  4.7622 -4.1437 -0.1684 -4.0939
#>  -5.9463  2.1020 -1.7775  1.0860 -0.7468
#>   0.2491 -0.8380  1.3709 -1.8596 -0.8215
#> 
#> (3,.,.) = 
#>   5.8651 -2.1167 -1.6930  2.1295 -0.2562
#>  -0.5099  0.6458 -1.3898  0.3735 -1.1454
#>   0.7393 -0.8239  2.7605  0.1086  3.5454
#> 
#> (4,.,.) = 
#>   0.7892  0.6603  1.0531  0.1704  2.2199
#>  -4.1030  3.3962  0.3652  0.8659  3.0989
#>   3.4354 -0.3811  2.5249 -1.5332  1.9719
#> 
#> (5,.,.) = 
#>   2.9805 -0.0562  0.3755 -0.9115 -0.1319
#>  -2.8097 -0.6731 -0.8753  1.2825 -0.5662
#>   1.6599 -0.2928 -0.5780 -0.1203 -0.8741
#> 
#> (6,.,.) = 
#>   0.4613 -3.6620  4.5320 -1.8052  1.8328
#>  -1.3925  4.3364 -0.0292  0.1149  2.6939
#>  -0.7831 -0.9873  1.5087 -0.6890  0.5075
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]