Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.7579 -0.7703  0.8588 -0.9332  0.3215
#>  -3.0967 -1.1644 -0.1485 -0.6196  0.5018
#>  -1.6448 -1.9527  2.7388 -0.5004 -1.2518
#> 
#> (2,.,.) = 
#>   0.5537  0.2613 -2.1762 -1.1287 -1.1530
#>   4.3006  3.4530 -2.7144  2.8584  2.1730
#>  -2.6771  1.0435 -1.7737  1.2404  4.1904
#> 
#> (3,.,.) = 
#>  -0.7432 -0.3940 -1.5890 -0.1633 -0.7355
#>   1.7391  1.7686 -0.7939  1.7135  1.8570
#>   2.2526  0.7933 -2.0034 -1.2825 -1.7115
#> 
#> (4,.,.) = 
#>  -3.5205 -0.0518 -2.4806 -1.2103  2.0199
#>  -0.1968 -0.5320  0.1065  0.2809 -0.7488
#>  -1.3248 -1.2188  4.0259 -0.1269  0.4840
#> 
#> (5,.,.) = 
#>  -0.8917 -0.6561 -1.0418 -1.3267 -1.1459
#>  -7.3179 -0.6640  1.7720 -0.2017  6.4245
#>   0.1682  0.8268 -1.7325 -1.5024  0.4245
#> 
#> (6,.,.) = 
#>  -1.0164  0.1472 -0.5471  0.0727  1.0642
#>  -0.1175  0.1261 -1.6407 -0.9427 -0.5131
#>  -2.1172  1.9714 -2.3723 -0.1332  4.9227
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]