Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   0.6115 -0.6945 -0.3265  0.4363 -0.9641
#>   1.2967 -2.9722  1.4547 -1.2541 -1.7555
#>   1.2006 -2.4482 -2.3533  0.4053 -1.5881
#> 
#> (2,.,.) = 
#>  -2.0045  1.7352 -0.6919  2.2877 -1.1013
#>  -0.0427  0.0462 -2.3348  4.0124 -3.8171
#>  -0.0302 -0.1257  0.5374 -1.8458  1.8937
#> 
#> (3,.,.) = 
#>  -1.2469  1.3874  0.9944  0.1205  0.5554
#>   0.2067 -3.3642  1.6143 -3.7644  0.7413
#>   2.4314 -4.2411 -0.1825 -0.9847 -2.4003
#> 
#> (4,.,.) = 
#>  -0.9246  2.8655  0.7092  0.7780  1.3078
#>   1.0962  0.5075 -0.2888  0.9480 -0.4185
#>  -1.5333 -2.6600 -0.9273 -0.1283 -2.1116
#> 
#> (5,.,.) = 
#>   0.0960 -1.3450  2.2790 -3.4128  1.8951
#>  -4.4611  5.2054  1.3082  2.4816  0.7197
#>  -5.5936  4.4032  2.3181 -0.4433  2.9499
#> 
#> (6,.,.) = 
#>   2.4518 -2.2804 -2.5193  1.5133 -2.5579
#>   0.2609  1.9061 -2.5038  4.4024 -2.4335
#>   0.7133  2.6330  1.5447 -0.2889  2.2180
#> 
#> (7,.,.) = 
#>   0.2977  1.1439 -0.5276  0.5689  0.6531
#>   6.0145 -6.4874 -1.5652 -2.0594 -2.1347
#>   1.7029 -2.2795 -1.5260  0.0385 -1.2926
#> 
#> (8,.,.) = 
#>   0.0073  0.7559  1.0937 -1.3279  1.8422
#>  -2.2426  3.3069  0.9948 -0.5853  3.1000
#>   0.5700  0.2864 -0.5471  0.5102 -0.0387
#> 
#> (9,.,.) = 
#>   3.4000 -3.2329  0.2287 -1.9157 -0.3814
#>  -2.4364  1.7518  0.8509 -0.1723  1.1750
#>   3.4129  0.4493  1.4372 -0.3006  0.5594
#> 
#> (10,.,.) = 
#>   5.3153  0.5438  0.4273  0.3986  0.4808
#>  -3.5701  5.5227  0.2550  2.8863  1.1066
#>  -1.5119 -1.7974 -0.4940 -0.9157 -0.5308
#> [ CPUFloatType{10,3,5} ]