Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.2670 -0.4704 -0.4694 -0.8322 -0.0362
#>  -1.2527  0.9000  2.0930  3.2794 -3.1562
#>   2.0033  0.5202 -1.8714 -3.3134  2.1068
#> 
#> (2,.,.) = 
#>  -1.1794  0.5252  1.8606  2.6199 -3.3844
#>  -0.4053 -0.7108 -0.0389  0.0956  0.3764
#>   1.7336 -1.8050  0.8208  1.1415  2.8968
#> 
#> (3,.,.) = 
#>   0.2489  1.4307 -0.5376 -0.3625  0.6273
#>   1.1675  2.4481 -0.7113 -0.8707  0.5959
#>  -1.0071 -4.5063  1.2991  1.7986  0.8913
#> 
#> (4,.,.) = 
#>   1.2894 -0.7181 -0.7021 -1.3536  1.8988
#>   0.9964  2.2473 -0.1325 -0.0535  0.0695
#>  -0.5818 -0.7045  0.6606  1.5254  0.6363
#> 
#> (5,.,.) = 
#>  -2.1512  3.2158 -0.8561 -0.6015 -3.0592
#>  -2.2999  1.3789 -1.2368 -1.5798 -2.6478
#>   0.7000 -1.1180 -0.2351 -0.9568  0.4439
#> 
#> (6,.,.) = 
#>   0.5878  3.9140 -1.2619 -1.5582 -0.5914
#>   1.4785 -0.1386 -1.0553 -1.2475  3.4167
#>  -0.0071  1.2966  1.9163  3.1992 -1.2905
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]