Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -4.6674 -1.4520 -5.2797  1.5998 -2.4838
#>   0.2830  0.6600 -1.6211  0.4874  0.3972
#>  -0.3477 -2.0802  0.7127 -0.3714 -1.3869
#> 
#> (2,.,.) = 
#>   0.6183  0.6643  1.5660 -0.3996  0.4283
#>  -0.7114  1.0525 -2.5011  0.8295  0.2734
#>   3.9196  1.9784  3.9913 -1.1467  2.3038
#> 
#> (3,.,.) = 
#>  -3.1729  1.3456 -1.2550  0.6792 -0.3995
#>   1.4022 -0.7748  3.9232 -1.2361  0.2528
#>  -2.6736  0.5437 -2.4289  0.8880 -0.6664
#> 
#> (4,.,.) = 
#>  -0.3992  1.9029 -1.3341  0.5764  0.6997
#>   0.7709  1.0706 -4.1820  1.1988  0.2825
#>   0.7846  0.7054 -2.1304  0.6021  0.2649
#> 
#> (5,.,.) = 
#>  -0.6055 -2.5206 -0.7939  0.0096 -1.5181
#>   3.8629 -0.0161  3.4608 -1.1902  1.0244
#>  -1.1990  1.7887  0.2802  0.1725  0.7868
#> 
#> (6,.,.) = 
#>  -0.7115  1.4012 -0.9233  0.4322  0.6169
#>  -1.7164 -0.7013  0.8697 -0.1964 -0.7181
#>  -0.6640  1.3453  2.1323 -0.4110  0.8727
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]