Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.2098  0.5153  0.2500 -0.7711  0.0070
#>  -1.8817 -0.1853 -3.8205  0.4326 -0.4455
#>   0.5470 -1.3113  0.2707  0.2071  1.1855
#> 
#> (2,.,.) = 
#>   4.6613 -0.0827 -1.7949  0.0302 -1.1464
#>  -2.6913  0.8166  5.7467 -0.7236  1.2034
#>  -0.5807 -0.8063  0.3280  1.8696 -0.3029
#> 
#> (3,.,.) = 
#>   2.1897 -0.1393  1.5328  0.9131 -0.6388
#>   1.3206  2.3480  1.4525 -2.0474 -1.0201
#>  -3.3564 -0.3619  1.3009 -0.4260  1.5596
#> 
#> (4,.,.) = 
#>   5.3612  1.0823  0.6923  0.8070 -2.6120
#>  -3.5739 -0.5444  3.0891 -0.3812  2.0865
#>  -1.7430  0.2183  1.5101 -0.7030  0.8731
#> 
#> (5,.,.) = 
#>  -2.9305 -1.5574 -3.6727 -0.7930  2.1018
#>  -2.3450  0.0171 -1.3270  0.8461 -0.4127
#>   5.5032  1.1741  0.7005 -0.2475 -1.9867
#> 
#> (6,.,.) = 
#>   1.3116 -1.3289  1.9941  0.9740  0.8273
#>  -0.8845 -0.7210  0.2710  0.1850  0.8491
#>  -6.1674 -1.5183 -0.7482  0.0738  2.5919
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]