Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   4.0840 -0.2669 -2.8872 -3.2387 -3.3035
#>   1.7170  0.2649 -1.1648 -2.4652  0.1504
#>  -1.7055  0.5796  1.8338  0.5635  0.3317
#> 
#> (2,.,.) = 
#>   0.4350  0.6016  0.3173 -2.7139  1.9286
#>  -0.2524  1.0583 -0.5244 -0.8770  2.1788
#>   2.6556  1.1123 -2.0264 -3.0577 -2.6177
#> 
#> (3,.,.) = 
#>  -0.6592  0.2023  0.9961  0.0003 -0.1153
#>   0.1471 -0.4395  0.2292  0.0326 -0.3614
#>   6.5016  1.0137 -5.4288 -6.5321 -3.1818
#> 
#> (4,.,.) = 
#>   2.7653  2.4890 -2.7888 -5.8919  2.3117
#>  -0.1289 -0.1775  0.0389 -0.5637  2.2680
#>   3.9144  1.8223 -4.1004 -5.2744  0.6297
#> 
#> (5,.,.) = 
#>   1.0076 -0.9213 -0.4954 -0.8649  0.8091
#>   1.1408  0.2914 -0.7599 -1.7538  0.0525
#>   0.9484  0.5327 -0.9273 -0.5100 -2.0599
#> 
#> (6,.,.) = 
#>   1.6623  0.8667 -1.2287 -1.4638 -3.2222
#>  -2.5809 -0.7970  1.9434  3.0636  1.7107
#>   0.6280 -0.3555 -0.7084 -1.2043  2.6593
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]