Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.3200  2.0860 -0.6598  0.6785  1.0456
#>   0.3707  0.9131  2.8244 -5.1988  3.7431
#>   0.6998  1.2414 -1.1083 -1.3115  1.3486
#> 
#> (2,.,.) = 
#>  -0.1074 -3.0098 -0.6100  1.3044 -2.0872
#>  -0.8024 -1.3482  1.0742  0.5429 -0.3499
#>   0.2216 -1.2165 -0.1257 -1.7994  1.1629
#> 
#> (3,.,.) = 
#>   0.5479 -1.4643 -0.7548 -0.8111 -0.2237
#>  -0.4026 -0.1296 -1.4261  1.3461  0.0213
#>  -0.3095 -0.4952  0.3468 -1.1304  1.3798
#> 
#> (4,.,.) = 
#>   0.9127 -0.6463  1.1742 -3.6200  1.4028
#>  -0.6729  0.4398  2.2256 -0.7352  0.8621
#>   0.0289 -1.3831 -0.6782  1.1443 -1.4597
#> 
#> (5,.,.) = 
#>   1.1011 -0.3013  0.7213 -3.5628  1.4120
#>  -0.3175 -1.0154  2.8197 -0.5319 -0.9188
#>  -0.9664 -0.7112  1.1948 -0.2910  1.0360
#> 
#> (6,.,.) = 
#>   0.9685 -1.0215 -2.4059  0.4187 -1.1646
#>   0.9270  2.4884 -2.1230  0.2040  0.3856
#>   0.9882 -0.1382 -2.9929 -0.2484  0.2288
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]