Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.3012 -1.1241  1.1523 -1.5602 -0.5858
#>   0.5090  1.1165  1.2415  1.5314  0.6055
#>   2.7698 -2.6363 -2.8351 -0.1183 -0.8260
#> 
#> (2,.,.) = 
#>  -3.0843 -1.6653  3.4495 -3.8033 -1.4476
#>   0.5454  2.1724 -0.5202  1.9395  1.0256
#>  -2.5867  0.3927 -1.7741 -2.2251 -0.1924
#> 
#> (3,.,.) = 
#>   0.6900  2.5669  9.4503  5.7568  2.1391
#>  -1.0166  2.3818  4.5624  2.5418  1.3926
#>   2.8692  3.7561 -7.1667  4.0435  2.0895
#> 
#> (4,.,.) = 
#>   1.3984  3.3005  4.1170  4.5757  1.8093
#>  -0.3123 -0.0971  1.1206 -0.4706 -0.2548
#>  -0.7372 -0.5234  0.2734 -0.7030 -0.1968
#> 
#> (5,.,.) = 
#>   1.0402 -0.1264 -3.0867  0.5539  0.2326
#>  -3.0396  0.3456  5.0636 -1.3179 -0.1766
#>  -3.6776  3.1207  2.4169 -0.8495  0.6628
#> 
#> (6,.,.) = 
#>   1.7868 -0.7309 -0.0768  1.6930  0.3017
#>  -0.8410  0.9162 -4.8901 -1.3685  0.0224
#>  -0.7855 -0.6515  2.5224 -0.8237 -0.4498
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]