Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   2.0991 -0.1046  0.6032  0.8438 -1.0748
#>  -2.9906 -1.5965 -0.1688 -0.8627  1.4513
#>   0.3210  0.8666  2.1716 -2.1679  1.1056
#> 
#> (2,.,.) = 
#>  -1.6983 -1.6506 -2.0359  0.9363 -0.4433
#>   4.3840  3.6166 -1.2635  0.7637 -2.7502
#>  -3.6113 -2.2526 -2.5333  0.0577  0.3243
#> 
#> (3,.,.) = 
#>  -0.4147  3.5511  0.7916  1.3028  2.1752
#>   0.0238 -0.7007 -1.0383  2.5846 -0.5191
#>  -2.8743 -2.0297  0.9226 -2.3903  1.6889
#> 
#> (4,.,.) = 
#>  -2.9209 -2.4688 -0.2762 -2.6624  0.7722
#>   0.4316  4.4723 -0.0936 -2.7827  0.6661
#>  -1.0489 -0.9685  1.5913 -1.5202  1.2450
#> 
#> (5,.,.) = 
#>   2.3258  3.4597 -0.3214  2.3340 -0.4382
#>  -2.5136 -1.2802  0.1538  1.7222  1.7999
#>   2.2579  2.5031 -1.4336  0.8873 -1.6317
#> 
#> (6,.,.) = 
#>   3.9082  3.5291 -0.9591 -1.1163 -2.5171
#>   2.9804 -0.1016  2.0931  0.9819 -0.7533
#>   1.3527 -0.6945 -1.2133 -0.3365 -2.0553
#> 
#> (7,.,.) = 
#>  -0.5113 -1.9690 -1.3068  3.4206 -0.5839
#>   1.5967  1.5237 -0.0254 -3.1562 -1.1767
#>  -1.5664 -1.2235  1.2893 -2.2818  1.2360
#> 
#> (8,.,.) = 
#>   1.8130  1.9611  0.9805 -2.7937 -0.4956
#>  -1.7759 -2.4418 -0.1561 -1.4492  0.1941
#>  -3.0896 -0.1488 -0.8940 -4.1825  1.0330
#> 
#> (9,.,.) = 
#>  -3.1269 -0.6206  0.4785 -0.8943  2.2812
#>   3.4966  4.7294 -2.3019  1.0470 -2.3488
#>  -2.8000 -4.9076  1.2158 -3.1347  0.7643
#> 
#> (10,.,.) = 
#>   1.7321  0.4194 -0.0089  2.7949 -0.7266
#>  -3.7125 -2.8770  1.9835 -4.0892  2.4518
#>  -2.3804 -0.2000  0.6451 -1.7811  1.8386
#> [ CPUFloatType{10,3,5} ]