Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   1.5828 -2.1540  5.5257  1.5698 -0.7564
#>  -0.3943 -1.6076  0.1886  1.0793 -1.8810
#>  -0.9112  2.2640 -1.8064 -1.3639  1.4482
#> 
#> (2,.,.) = 
#>   0.9566 -1.3446  2.1954  0.9686 -0.2724
#>  -1.7092  3.9689 -1.8511 -3.7171 -0.4236
#>   0.0308 -2.8527  1.5768  3.2814 -0.0617
#> 
#> (3,.,.) = 
#>  -0.0240  0.7274 -2.5283 -0.7940  0.5197
#>   2.8053 -4.6622  2.8299  3.1261 -1.0330
#>   4.0411 -2.6250  8.4744  1.4893  0.5249
#> 
#> (4,.,.) = 
#>   0.8720 -0.4780  2.1447 -0.2738 -0.9081
#>   0.6510 -0.2255 -1.8818 -0.9528 -1.1133
#>   0.4828  1.1404 -0.8155 -0.8593  1.5481
#> 
#> (5,.,.) = 
#>  -1.2668  0.8188 -3.5250 -0.7091 -0.4585
#>   0.1879 -0.5187  1.7766  0.7461  0.1929
#>   1.0676 -1.6145 -4.3234  0.5936 -0.0767
#> 
#> (6,.,.) = 
#>   1.6051 -1.8161 -0.0025  1.4389  0.8885
#>  -0.1828  3.1214  1.9642 -2.2188  1.7534
#>  -0.7339  3.5144 -0.8780 -4.2507 -1.4094
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]