Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   0.9021 -0.2303 -1.9573 -0.4295  0.4943
#>   1.2040 -0.5276 -3.7314 -0.4447 -0.0137
#>   1.9537  0.5243 -3.6375 -0.0013 -1.3906
#> 
#> (2,.,.) = 
#>   0.1541  0.7863 -0.8593  2.7808 -0.9805
#>   1.4967 -0.6461 -4.0598 -1.8737 -0.7285
#>   1.2912 -0.2154 -2.2134 -1.6077 -0.0560
#> 
#> (3,.,.) = 
#>  -0.2907  0.0193  0.3479  0.5718  0.1546
#>  -1.2076  0.0154  1.2007  1.6105 -0.3953
#>  -0.2020 -0.6206 -1.8230  0.4767 -0.3117
#> 
#> (4,.,.) = 
#>  -1.7208  0.0689  3.7835  0.7450  0.7000
#>   0.1810 -0.1306 -0.4503 -0.1270  0.3326
#>   0.0201 -0.1265  2.1628 -1.9269  1.7956
#> 
#> (5,.,.) = 
#>  -0.9338  0.2446  2.5532  0.2852  0.1203
#>  -2.5259  0.1202  3.3482  3.1141 -0.1742
#>   1.2186  0.4413 -2.3006  0.1435 -1.2534
#> 
#> (6,.,.) = 
#>   0.4723 -0.9953 -3.2615 -0.4619  0.2510
#>   0.4925  0.0812 -2.9198  1.4771 -1.5713
#>   1.2868 -0.3890 -3.1591 -0.8692  0.0690
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]