Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   0.1484 -0.9835 -1.7588  0.2188 -3.0843
#>   0.6351 -3.1357 -0.0008  6.2433 -1.1671
#>   0.9797 -4.8722 -1.3975  2.6581  0.8835
#> 
#> (2,.,.) = 
#>  -0.5620  0.4367  0.3048  2.9712 -0.7842
#>  -1.3895  5.1028  0.4180 -1.8640 -3.0902
#>  -0.3480  0.3246  0.6724  4.4703 -1.6923
#> 
#> (3,.,.) = 
#>  -0.3891  3.1541  1.3916 -0.9026 -0.9251
#>  -0.3860  0.6418 -0.2080 -0.3345 -0.2160
#>   0.4306 -1.3080 -0.3916 -2.8263  2.7174
#> 
#> (4,.,.) = 
#>  -1.9861  7.3143  1.2850 -2.8626 -2.5561
#>  -1.0059  2.6421  0.0879  2.6169 -4.0650
#>   0.6414 -2.5179 -0.5408 -1.7255  3.1612
#> 
#> (5,.,.) = 
#>  -0.4087 -0.5196  0.4274  0.0310  3.5647
#>  -1.0305  0.5722  0.3101 -1.3338  4.3689
#>   2.3748 -6.7450 -1.9669  2.6457 -1.5191
#> 
#> (6,.,.) = 
#>   2.1195 -6.9232 -1.0330  5.4318 -0.3499
#>  -1.3679  4.9450  0.1247 -3.0418 -2.5399
#>  -0.4706 -1.0621 -0.1316 -0.2671  3.7693
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]