Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.5377 -2.0946  1.9283 -1.0583  0.5486
#>  -2.0610  0.9258 -1.3998  0.0679  4.0662
#>  -1.7478 -1.1335 -1.3330 -0.0247  1.1185
#> 
#> (2,.,.) = 
#>   0.6247 -0.1615 -0.0545 -1.2726  1.5450
#>  -0.3370  3.1739 -2.2987  1.9350 -2.0867
#>  -2.2894  0.4201  0.6217 -0.1967  6.3548
#> 
#> (3,.,.) = 
#>  -0.0126  0.1456 -0.4895 -0.4161  0.7882
#>  -0.2956  0.4910  0.8241 -0.2723  2.6308
#>   1.1750 -0.1503  1.9447  0.4291 -1.9672
#> 
#> (4,.,.) = 
#>  -2.6704 -0.4561 -2.6133 -0.5652  4.0791
#>   0.1057  1.1978  0.3418  0.5246  0.3017
#>   0.0553 -2.0565 -0.3674 -0.5168 -1.7554
#> 
#> (5,.,.) = 
#>  -0.5611 -2.0888  0.5812 -1.2372  2.0965
#>   0.2226  0.3081  0.0692  0.0380 -0.1120
#>   0.2746  0.4811  0.1463  0.3549 -0.7090
#> 
#> (6,.,.) = 
#>   0.2136 -1.1365  0.9610 -0.3923 -0.0582
#>  -1.7842 -0.0751 -0.9920 -0.0095  2.7852
#>   0.8289 -1.8879  1.7930 -0.7551 -0.6629
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]