Matmul

Matmul

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.3145 -0.5595 -0.0690  0.5759  0.2661
#>  -0.3851  2.6199 -3.1328  2.8422 -0.2790
#>  -2.3893 -2.2760  1.5882 -3.0124 -0.5051
#> 
#> (2,.,.) = 
#>  -0.7015 -0.4270 -0.1721  2.1045 -0.6880
#>   1.3278  1.0885 -1.2500  1.7052  0.9912
#>  -0.8212 -1.0197 -0.7320  4.4162 -0.1785
#> 
#> (3,.,.) = 
#>   1.1285  1.3519 -1.6765  1.4334  1.1796
#>   0.3272  3.0396 -2.9070  3.4524 -0.6724
#>   0.2266  1.0626 -0.0078 -0.1805 -1.0016
#> 
#> (4,.,.) = 
#>  -0.8478 -0.7994  1.0614 -3.1091 -0.1066
#>   1.0050  1.7832 -2.5138  4.0749  0.7012
#>  -0.6469 -2.1022  1.4766 -1.2548  0.3860
#> 
#> (5,.,.) = 
#>  -1.8415 -1.5547  0.3677 -0.7214 -0.0520
#>   2.3493  4.8668 -2.6966 -0.1245  0.0409
#>   1.0042 -0.8043  1.9781 -2.0490 -0.0557
#> 
#> (6,.,.) = 
#>  -0.7894 -0.9777  0.6695 -0.3267 -0.3456
#>   0.7919  3.9030 -3.8567  3.2025  0.1171
#>  -1.4926 -1.7354  0.7417  1.7177 -0.9489
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]