Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.7507 -0.0873  0.5344 -1.5103  0.8393
#>  -2.7561 -0.7660 -1.1228 -4.9524  3.4177
#>  -1.8057 -0.3449  0.2942  1.4129 -0.1947
#> 
#> (2,.,.) = 
#>  -1.0385 -0.5117  0.7700  0.6451  0.9210
#>   1.0159  0.4831 -0.0682  2.6191 -2.3171
#>  -0.8033 -0.1980 -2.3872 -1.0535 -0.0150
#> 
#> (3,.,.) = 
#>   0.3240  0.1687  1.3306 -0.9834  0.6722
#>  -1.0368 -0.9682  0.7735 -1.8243  3.2775
#>  -0.0791 -1.3968 -1.4155  1.9246  1.7017
#> 
#> (4,.,.) = 
#>   0.4893 -0.6543 -1.9460 -2.0980  2.0102
#>   0.4588 -0.8679 -1.4432  0.2272  1.4623
#>   1.3707  0.5524 -0.9724 -0.4408 -1.1298
#> 
#> (5,.,.) = 
#>   0.0408  0.3578  2.0797 -0.4886  0.1892
#>   2.2815 -0.5862 -1.2594  1.3359  0.6577
#>   0.1539  0.2445  0.7765 -0.6007  0.0605
#> 
#> (6,.,.) = 
#>  -0.3843  0.1339 -0.9178  0.0377 -0.7403
#>   0.9258  1.0118 -4.5009 -1.5748 -2.9708
#>  -1.7505 -0.2612 -0.7348 -2.9377  1.5311
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]