Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.9167  1.1517 -0.5897 -0.2039 -1.0299
#>  -0.3668 -0.9138 -0.9727  0.7339  0.5669
#>  -1.2136  0.2873 -3.9519  1.6645 -0.0539
#> 
#> (2,.,.) = 
#>  -1.1613  1.6900 -2.8405  2.6566 -1.4254
#>  -0.8980 -2.0666 -1.2913  2.9247  0.8088
#>   2.4297 -0.2475  3.6336 -2.1976  0.7700
#> 
#> (3,.,.) = 
#>   0.9605 -1.1265  0.2791  0.9571  0.9848
#>   3.1733 -0.3693  1.0726  1.4184  1.2385
#>  -0.0560  0.3121 -3.1315  1.2113  0.3454
#> 
#> (4,.,.) = 
#>  -1.9138  1.6272  0.2749 -3.8130 -1.4323
#>   1.8724 -0.6448  1.4162 -0.7338  1.0873
#>   2.2765  0.5757  1.5940  1.4776  0.0345
#> 
#> (5,.,.) = 
#>  -3.2278  2.4879 -4.1677  3.1417 -2.6885
#>   0.4410  1.3773 -1.3865  0.7383 -0.5008
#>   0.8000 -1.1504  1.9837 -1.7440  0.9514
#> 
#> (6,.,.) = 
#>  -3.1205 -1.9122 -1.3471  1.3484 -0.0646
#>  -1.4873 -1.5667 -0.6279  0.3892  0.4522
#>   1.1187 -0.4504  1.1803 -2.8332  1.0127
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]