Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   2.2212 -3.7360  3.6128  4.3205  1.7180
#>  -0.8522 -0.1409  2.3900 -0.0118 -0.2700
#>   5.2143 -0.1304 -6.7105 -0.8499  0.1775
#> 
#> (2,.,.) = 
#>   3.4524  0.1249 -5.0345 -1.9593 -0.7891
#>  -0.6715  4.4294 -4.1720  0.1468  1.9942
#>  -5.3132  4.3873  1.8608 -3.0499 -1.1165
#> 
#> (3,.,.) = 
#>  -4.6296  1.6081  4.0484 -0.0456 -0.0585
#>   1.3100 -0.2991 -1.7727  1.7964  1.4130
#>   0.3922  4.6351 -4.5374  1.7541  3.3697
#> 
#> (4,.,.) = 
#>   1.0989  0.3837 -0.2755  2.0519  1.7687
#>   1.8521 -3.9821  1.8916 -0.6304 -1.9404
#>  -3.2138 -0.3769  4.5113  0.7756 -0.1240
#> 
#> (5,.,.) = 
#>   4.1195  4.1837 -10.0723 -1.7745  1.2612
#>  -0.8389  0.8251 -0.0623 -0.7531 -0.3042
#>  -2.1405  2.0301  1.4868  1.0605  1.2867
#> 
#> (6,.,.) = 
#>   1.7165 -0.1520 -1.6944  0.5225  0.5750
#>  -0.5101 -0.9448  1.4539  1.3430  0.4865
#>  -3.3028  6.3948 -3.9811 -2.2649  0.7398
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]