Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -5.1641  2.8806 -2.6738 -0.5202  2.5728
#>   2.2125 -2.8874  0.0337 -1.6516 -2.7068
#>  -3.2578  0.4841 -2.7980 -1.8945  2.1466
#> 
#> (2,.,.) = 
#>  -0.1574 -0.3310 -0.6516 -0.3010  1.2618
#>  -0.2616  0.9704  1.2957 -0.4373 -1.4102
#>   1.1745 -0.8641  0.4495 -0.3805  0.1986
#> 
#> (3,.,.) = 
#>  -3.6540  5.4509  1.3682  1.9462  2.7942
#>  -1.0043 -3.1312 -3.3738 -4.1573 -0.6638
#>   0.0188  2.5586  2.1171  2.4287  1.1574
#> 
#> (4,.,.) = 
#>  -0.9699  0.5037 -0.6680 -0.0429  1.2082
#>  -0.6497  4.2699  3.1770  3.0852  1.4933
#>   0.3205  2.2010  2.2246  2.0913  0.7186
#> 
#> (5,.,.) = 
#>   1.9751 -0.0294  2.0492  0.9070 -0.6322
#>  -2.9091  4.0232  0.6975  1.5305  2.1840
#>   1.7337 -2.3486 -0.1972 -1.5010 -0.5846
#> 
#> (6,.,.) = 
#>  -2.6074  2.5011 -0.6440  0.6382  3.2248
#>  -2.8262  3.2865 -0.3437  1.5613  3.6004
#>   1.3922 -6.2041 -3.9366 -5.3273 -0.4870
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]