Matmul

Matmul

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.7675  0.3379 -0.5562 -1.0605  1.2292
#>  -1.5549 -1.5068 -1.6701 -2.4790 -1.6464
#>   1.5212  1.3090  0.8613  0.8281  0.4620
#> 
#> (2,.,.) = 
#>   0.2340  0.7120 -0.1111 -1.9591 -0.6993
#>  -0.8983 -1.5865  0.0452  2.0367  0.0378
#>  -0.7341  2.7239  0.6071 -2.8259  3.1268
#> 
#> (3,.,.) = 
#>   3.8140  2.5278  1.7050  3.7723  1.7137
#>  -0.1898 -0.3299 -1.2124 -3.6669 -2.7894
#>  -0.0936  0.0576  0.1754  0.0217  0.1043
#> 
#> (4,.,.) = 
#>   0.8587  0.3662  1.3822  1.6207 -0.1883
#>   0.0054  1.9542  0.3394 -4.5611 -0.8389
#>  -0.7280 -1.6662 -1.9565 -0.0591 -0.5331
#> 
#> (5,.,.) = 
#>   1.2511  2.6787  1.9294  0.7742  2.5080
#>   0.0167 -1.2511 -0.5554  2.6734  0.5818
#>  -2.1661 -2.6419 -2.5364 -1.8706 -1.6541
#> 
#> (6,.,.) = 
#>   1.2740  2.5813  1.7152  1.1671  2.8568
#>  -2.3233 -1.5228 -1.7941 -1.7845  0.1730
#>  -0.2156 -2.5358 -0.8445  1.4851 -2.4990
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]