Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -2.0038  1.5230  0.5681  0.0927  0.5975
#>  -2.8164  0.3671  0.4627  0.6987 -0.4495
#>  -2.1566  1.3466  0.1475 -1.2630 -0.5913
#> 
#> (2,.,.) = 
#>  -3.2801 -0.3992 -0.0754  1.4805 -0.9787
#>   0.8145  0.9914 -0.0831 -0.9736  0.6240
#>   4.0725 -1.1388 -0.7370 -0.5983  0.3570
#> 
#> (3,.,.) = 
#>   0.3763  2.0305  0.1416 -4.5523 -0.8800
#>   2.3155  0.3074  0.6006  2.1991  2.9026
#>   1.3539  0.7974  0.7434  2.3823  3.0331
#> 
#> (4,.,.) = 
#>   1.7234  0.0061 -0.0808 -1.4487 -0.1204
#>   2.0586 -1.6249  0.0224 -0.7228 -0.9207
#>  -0.5356  0.1881  0.0585 -2.2608 -1.5066
#> 
#> (5,.,.) = 
#>  -5.3128 -0.1730  0.4157  2.8229 -0.7681
#>  -0.4694  1.2870  0.4609 -4.3323 -1.7211
#>   0.5327 -1.9926 -1.1536 -0.5509 -2.1372
#> 
#> (6,.,.) = 
#>   0.6566  1.7562  0.1222 -1.4541  0.9705
#>  -2.5228  1.1097  0.1487 -0.8086 -0.6850
#>   1.4285  0.4281 -0.9629 -3.9915 -1.7098
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]