Matmul

Matmul

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   1.8786  0.7805  0.5902  1.5477  0.1583
#>  -2.8090  0.3525 -1.2957 -1.3978 -0.1770
#>  -0.2576  0.6066  0.0541 -1.5795  2.2743
#> 
#> (2,.,.) = 
#>   1.9499 -0.0318 -0.8813  1.7104 -0.9636
#>   6.5289  1.3875  0.4037  4.6400  0.0217
#>  -4.8476 -0.7085  0.1826 -2.1623 -1.2229
#> 
#> (3,.,.) = 
#>  -4.8888 -1.0743 -0.1955 -3.6174  0.1509
#>   3.6413 -0.2861 -0.4438  3.5767 -2.2158
#>  -0.8250  0.3291  1.2735 -1.2568  1.5041
#> 
#> (4,.,.) = 
#>   2.3035  0.5589 -0.8020  2.4206 -1.0837
#>  -2.1565 -0.8847  2.4436  0.7768 -2.6741
#>  -2.6241 -0.1260  0.7893 -3.0718  2.0199
#> 
#> (5,.,.) = 
#>   1.1707  0.7229  0.8568 -0.7896  2.5422
#>  -0.4785  0.3025  1.6764 -0.8939  1.3555
#>  -0.0063 -0.7727 -0.2007 -0.9560  0.4830
#> 
#> (6,.,.) = 
#>  -0.7734  0.2588  1.5157 -2.5587  3.1241
#>  -1.0703  0.6214  0.6939 -1.7583  2.0776
#>  -3.1887 -0.5154 -0.3526 -0.6870 -1.8410
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]