Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.3776 -1.5699 -0.5134 -1.2103 -1.0586
#>  -1.4508 -2.0213 -0.1480 -0.8909 -2.2765
#>  -0.9439 -1.0493 -0.1309 -0.0080 -1.5252
#> 
#> (2,.,.) = 
#>  -0.8822 -0.9502  1.4048 -2.8583  0.1733
#>   1.0048 -1.3646 -0.1712  1.5967  1.3787
#>   0.5894  2.1189  0.9424  0.1588  1.8092
#> 
#> (3,.,.) = 
#>  -2.0221 -0.8978  0.9612 -0.7794 -2.0082
#>  -2.4678 -0.9668  1.8414 -2.5926 -1.7553
#>   0.0835 -0.8381  0.9191 -0.6654  1.1165
#> 
#> (4,.,.) = 
#>   1.7683  1.1440 -0.7631  0.4144  1.8258
#>  -0.2301 -2.9295 -0.0376 -0.0855 -0.2925
#>  -0.3743 -0.8254 -0.2772 -1.4743 -0.8297
#> 
#> (5,.,.) = 
#>   1.1734 -0.1834  0.3996  0.9801  2.1932
#>   1.6206  0.7471  0.7955  0.6952  3.2521
#>   0.9458  2.5929 -1.0958  0.3742  0.1917
#> 
#> (6,.,.) = 
#>  -0.0993 -0.9410  0.1755  1.2893  0.0660
#>  -0.1822 -0.7323 -0.1036  2.2127 -0.3575
#>   0.7232  0.9718  0.6476  0.8768  1.7385
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]