Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   2.9185  0.3479 -2.0819 -1.8383 -4.4007
#>  -1.6705  0.6004  0.9518  1.3339  2.9970
#>   0.3537 -0.0651 -0.1409  0.0905 -0.3957
#> 
#> (2,.,.) = 
#>   0.8310  1.2702 -0.3669  2.3671  0.4599
#>  -2.2166  0.5343  1.3758  1.4164  1.5188
#>   0.2989  0.1971 -0.1437  0.9404  3.1282
#> 
#> (3,.,.) = 
#>   2.1368  1.8933 -1.4111  1.5963 -2.5206
#>   0.6988 -0.7572 -0.4765 -1.2103  1.0179
#>   1.4800  0.1518 -1.0842 -0.8377 -0.8209
#> 
#> (4,.,.) = 
#>  -1.3366 -0.3241  0.9642  0.6252  1.8930
#>   1.9990 -0.8803 -1.2821 -2.2908 -2.9463
#>  -1.3627 -0.7532  0.7391 -1.0635  1.2541
#> 
#> (5,.,.) = 
#>  -0.6787  0.3050  0.3201  0.3352  1.0467
#>  -0.2514 -1.6554  0.1880 -2.0598  0.9453
#>  -1.5042  0.3858  0.7643 -0.0260 -0.9697
#> 
#> (6,.,.) = 
#>   0.6168  1.2695 -0.1389  2.6813 -0.1228
#>  -1.6744  0.3343  1.1393  1.9707  4.6329
#>   1.8991 -0.1984 -1.5494 -3.2538 -6.7267
#> 
#> (7,.,.) = 
#>   0.8955 -0.2154 -0.5586 -0.4097  0.3897
#>   0.5789 -0.3816 -0.0985 -0.0300 -2.4265
#>   0.5199  0.3187 -0.3724 -0.0589 -1.3431
#> 
#> (8,.,.) = 
#>  -0.5801  0.2319  0.2399  0.0362  0.4656
#>  -0.1021  1.0653 -0.2086  0.5247  0.3005
#>  -0.9609 -0.2868  0.2830 -1.2864  1.2329
#> 
#> (9,.,.) = 
#>   0.7411  0.9152 -0.2739  2.0552  0.6770
#>   1.4019  1.4527 -1.0964  0.6220 -1.9376
#>   0.1877 -0.0084  0.0030  0.1538 -1.6132
#> 
#> (10,.,.) = 
#>   0.4974 -1.8509 -0.2767 -2.7295 -0.5511
#>   1.9495  1.0476 -1.2585  0.3402 -3.7378
#>   1.0375  0.0991 -0.5563  0.2654 -0.2848
#> [ CPUFloatType{10,3,5} ]