Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>   2.9089  0.9845 -1.9209  0.7565 -0.8089
#>   1.5602 -0.1355 -0.5663 -0.7256 -0.3005
#>  -5.1681  0.8439 -0.2311  4.0118  0.6029
#> 
#> (2,.,.) = 
#>  -3.2413  0.6460 -0.0018  2.4108 -0.3839
#>   1.1368  0.6532  1.2959 -0.8586 -1.9191
#>   1.8520 -0.0187  0.1186 -0.9161  0.4118
#> 
#> (3,.,.) = 
#>   2.9787  0.2050  0.3688 -1.8496 -1.7943
#>  -2.8292  0.7724  0.4590  2.3608  0.5896
#>   2.1940  0.3648  0.9281 -1.0902  0.0810
#> 
#> (4,.,.) = 
#>  -2.1130  0.3511 -0.6688  2.3332  1.7527
#>   3.4641  0.3747  1.2832 -2.0061 -0.3228
#>   3.3574 -0.2530  0.2699 -2.4027 -0.9421
#> 
#> (5,.,.) = 
#>  -0.7329 -0.3116 -0.9270  0.1775 -1.1252
#>  -0.5466  0.5188 -0.3335  1.2077  0.1667
#>  -2.5609 -0.2285 -1.5218  2.0231  0.8828
#> 
#> (6,.,.) = 
#>  -2.6296  1.0644 -0.3949  3.1285  0.5754
#>   0.6244 -1.2722  0.0847 -2.0712  0.1138
#>  -1.6715  0.1444 -2.2961  2.2244 -0.3139
#> 
#> (7,.,.) = 
#>   2.3106 -1.2625  1.1150 -3.0800  1.7108
#>  -1.4338  0.6178  0.9726  1.0354  0.0449
#>   0.5690 -1.0606 -0.5211 -1.1207  1.3386
#> 
#> (8,.,.) = 
#>  -0.7904  0.7064  1.5994  0.5113  0.1367
#>  -0.6828 -0.1212 -2.8078  1.8834  0.6446
#>   1.6149 -0.3797 -0.5884 -0.9357  0.2698
#> 
#> (9,.,.) = 
#>   2.2742 -1.5073 -0.8308 -2.4203  1.3077
#>  -3.3323 -1.0881 -2.3815  2.2911  3.2742
#>  -1.8565 -0.3452 -0.5353  0.8325  0.3332
#> 
#> (10,.,.) = 
#>   2.3585 -0.4239  0.6924 -2.1180  0.0934
#>   0.9073  1.1976 -0.0628  0.8384 -1.5615
#>  -1.1092 -0.2548  0.6868 -0.3382 -0.6751
#> [ CPUFloatType{10,3,5} ]