Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.1565  0.4495 -0.6354 -0.3234  0.7672
#>   0.6624 -0.4056  0.6719 -0.1669 -1.6207
#>   1.0852  1.1319 -2.3298  0.2497 -0.9297
#> 
#> (2,.,.) = 
#>   1.1429  3.0684 -5.9933 -3.2019  0.7114
#>   0.5748  0.5042  0.8153 -0.1866 -0.0380
#>  -0.2703  1.4804 -1.6916  0.0900  2.2382
#> 
#> (3,.,.) = 
#>   5.9538  5.2667 -2.6228 -1.2204 -3.5945
#>  -2.6554 -3.5252  2.6982  0.3088  0.2632
#>  -0.5706  0.0646 -0.2927 -2.1494  0.9773
#> 
#> (4,.,.) = 
#>   3.2612  2.7454  1.3971  0.1264 -1.3139
#>   2.0284  0.1369 -0.1128 -1.3786 -3.5403
#>   0.3078  0.2328 -2.0552  0.3774 -0.8027
#> 
#> (5,.,.) = 
#>  -0.8434  0.3623 -0.3323 -1.7928  1.9226
#>  -0.2445  0.9932  0.0609 -1.1827  1.9326
#>  -0.7318 -1.3635  1.0100  0.0545 -0.4447
#> 
#> (6,.,.) = 
#>  -1.6605 -0.2965  1.0259 -0.6921  2.8388
#>   0.3025 -2.1373  4.2295  2.3333 -2.4647
#>  -2.3783 -4.3968  2.1835 -2.4612 -1.8016
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]