Matmul

Matmul

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.1864  0.1016 -1.3007 -0.9707 -0.3342
#>   1.0613 -1.5288  0.7496 -1.9112 -0.8753
#>   0.5912 -0.3715  0.5712 -2.0003 -0.6216
#> 
#> (2,.,.) = 
#>   1.1324 -4.2583  0.4448  1.9275  1.2737
#>  -0.1222  1.5433  1.3553 -1.1332  0.3043
#>  -0.1790  0.3083 -0.2293  0.5862 -0.2805
#> 
#> (3,.,.) = 
#>   0.8300 -2.3243 -0.1491 -1.1395  0.3158
#>   0.9951 -1.4299  0.7763 -1.6064 -0.8442
#>  -0.6136  0.5805  0.6603  3.1572  1.3482
#> 
#> (4,.,.) = 
#>   1.0351 -2.2652 -1.3221 -2.5933 -1.7704
#>  -0.1478  1.0040  2.0664  2.6581  0.0679
#>   0.2997  1.4344  1.7470  0.4908 -2.5126
#> 
#> (5,.,.) = 
#>  -0.1510  0.9140  1.0146 -0.3200  0.7257
#>   1.4374 -3.0519  0.5056 -2.5645  0.0626
#>  -1.3307  0.8666 -1.6609  3.7003  1.3258
#> 
#> (6,.,.) = 
#>   0.5835 -1.6227 -0.4386 -0.1676 -0.7077
#>   0.1231  0.4861 -0.9410 -2.6570 -1.3280
#>  -0.4508  2.3568  1.3555  0.4188 -0.4210
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]