Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -2.6809 -0.4946 -10.5166  4.8508 -0.9662
#>   0.2991 -1.0008 -2.9917 -3.6978 -3.4106
#>   1.1409  0.2604  2.9629 -3.1385 -1.4678
#> 
#> (2,.,.) = 
#>   0.8839 -1.4735  0.0652  0.7135  0.8053
#>  -0.8732  0.0161 -1.2564 -1.8175 -0.2341
#>  -0.5970 -3.7456 -4.1767 -0.8696  2.6182
#> 
#> (3,.,.) = 
#>   0.1990  2.0418 -1.1464  5.1929 -2.0088
#>  -0.3292  0.0830  0.3383  1.6647  1.5766
#>   0.2309 -1.2398 -1.2352 -1.9865 -0.6188
#> 
#> (4,.,.) = 
#>   -0.0906  -0.9770  -4.1810  -0.3416  -2.1877
#>    3.6772   0.1005  12.2702  -2.9408   1.9869
#>    0.9927   1.5519   1.8898   4.2136  -0.6510
#> 
#> (5,.,.) = 
#>  -1.6850  0.7892 -0.7554 -2.1070  0.5298
#>  -0.7920 -3.8116 -3.9629 -5.6347  1.1118
#>  -2.1208  1.5032 -4.8464  4.3452 -0.2817
#> 
#> (6,.,.) = 
#>   0.5870 -1.1692 -1.5098  2.6025  0.4319
#>  -1.8684  1.0197 -3.8004  2.1111 -0.1764
#>  -0.5343  0.2543 -5.2417  5.1757 -1.5291
#> 
#> (7,.,.) = 
#>   0.2717  0.9324  3.0536 -2.0161 -0.1574
#>   0.5406 -0.0661 -0.9495  2.3033 -0.7396
#>   1.4107  0.3536  2.9462  2.5392  0.4394
#> 
#> (8,.,.) = 
#>  -1.7402  2.0616 -3.0313  1.0960 -1.7451
#>  -0.6110 -2.9992 -7.7965  5.3914  1.2558
#>   1.5597  0.2508  4.2138 -3.1801 -1.2046
#> 
#> (9,.,.) = 
#>  -0.0096  1.2711  0.1790  0.4257 -1.4911
#>   0.3910  0.7790  0.8669  1.2045 -0.6025
#>   0.6989  1.7260  2.6494  1.0718 -1.1003
#> 
#> (10,.,.) = 
#>  -0.2817 -2.0163 -4.4983  5.0457  1.8857
#>  -0.3029  0.9321  1.4816  3.8409  2.2694
#>   0.6240 -1.3195 -1.5015  0.9746 -0.1551
#> [ CPUFloatType{10,3,5} ]