Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   0.6041  0.6619 -0.6371 -1.0380 -2.0770
#>   1.7441  2.0100  0.7980  0.2016  2.9895
#>  -2.9781 -0.5244 -1.3785 -1.8876 -3.8417
#> 
#> (2,.,.) = 
#>   3.0330 -1.5650 -0.0510  0.2147 -2.5532
#>   1.9785  1.0165 -0.8439 -1.1079 -3.0210
#>  -0.5486 -1.1823  0.7996 -2.1450  0.2680
#> 
#> (3,.,.) = 
#>   1.0651 -0.9382  2.2542  0.2489  5.0051
#>  -0.5103  0.4860 -0.6954 -0.3941 -1.4967
#>   3.2346 -0.3184 -0.9384  0.4963 -3.9229
#> 
#> (4,.,.) = 
#>   1.4621 -1.0502  0.1900 -0.6354 -1.3026
#>  -0.4788 -1.5179  1.1752 -2.5428  0.7629
#>   1.4806  1.5572 -0.1408  2.2100  1.4014
#> 
#> (5,.,.) = 
#>  -2.1577  0.4691 -0.2859 -0.0262  0.5391
#>   2.2151 -2.2255  0.5550 -1.4596 -2.0202
#>  -1.5580  1.8312 -0.6325  0.2425  0.5216
#> 
#> (6,.,.) = 
#>   0.8914 -1.0515  0.1855  0.1650 -0.6000
#>  -0.8602  0.9691  0.4082  1.3578  2.9909
#>  -1.3983 -0.6668  0.8941 -0.8310  2.0455
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]