Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.4466  2.8044  2.8174  2.5041  5.2175
#>   0.2367  1.0715  0.5079  0.8159  2.2284
#>   0.1630 -2.0008  2.5111  1.0549 -4.8251
#> 
#> (2,.,.) = 
#>  -0.4308  0.0212 -0.2369 -0.8184 -0.2933
#>   0.2498 -1.4315 -1.8357 -0.6454 -2.1434
#>  -2.9289  0.8148  1.8896 -4.1475 -0.6263
#> 
#> (3,.,.) = 
#>  -1.3705  3.7351  4.0344  0.0502  5.5816
#>  -1.4132  2.7273  2.0054 -1.1071  4.1481
#>   0.2536 -1.1930 -0.8290 -0.1364 -2.0262
#> 
#> (4,.,.) = 
#>  -0.3421  1.0003  0.1416 -0.3518  1.7703
#>   0.1171 -0.8847  0.6646  0.3339 -1.9094
#>  -0.9799  0.4180  0.5100 -1.4405  0.2303
#> 
#> (5,.,.) = 
#>   1.9658 -1.2240  1.1241  3.7031 -1.6190
#>  -0.7389  0.0966  2.4269 -0.2066 -0.9655
#>   3.0809 -1.9082 -2.5369  3.9285 -1.2995
#> 
#> (6,.,.) = 
#>   0.7245  1.4538 -3.1329  0.1243  4.4277
#>   2.2290 -1.7824 -2.7336  2.3748 -1.4900
#>  -1.2957 -0.5611  0.3263 -2.2047 -2.0847
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]