Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   0.2301  1.1657 -1.5565 -2.0733  0.6698
#>  -3.0178  3.3180 -0.5570 -3.3573  2.1463
#>  -1.2832  1.8379 -2.0699 -1.9677  0.0243
#> 
#> (2,.,.) = 
#>  -0.6477 -1.2312  3.2865  0.4747  0.9384
#>   0.3544 -1.8473  4.1643  0.8304  1.4965
#>  -1.5859  0.9722 -0.3948 -2.2738  0.4639
#> 
#> (3,.,.) = 
#>   0.0316 -0.1523  0.1210  1.2424 -0.4314
#>   1.6589 -0.2982 -1.5107  3.5818 -1.7249
#>   2.7293 -3.5496  1.7617  2.4844 -1.2821
#> 
#> (4,.,.) = 
#>   1.1869 -1.7780  1.7616  3.9081 -0.8921
#>   1.0573  0.7437 -2.3689 -3.2055  0.4360
#>  -1.1171  2.3065 -2.2266 -3.1266  0.8167
#> 
#> (5,.,.) = 
#>  -1.1221  0.5445  0.4752 -2.9302  1.2214
#>   2.0789 -1.5914 -1.9775  5.3311 -3.6772
#>  -1.2056  0.1955  1.7159 -1.6529  1.4248
#> 
#> (6,.,.) = 
#>  -3.7053  2.4316  1.2503 -3.6938  2.4663
#>  -0.6434 -0.8126  1.9358 -0.2965  0.5066
#>  -2.1355  2.5167 -1.8817 -3.0724  0.7091
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]