Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   0.2204  0.0976 -0.6607  0.0946  0.2527
#>   0.7142 -0.8612 -2.1390 -5.7182 -0.2748
#>   0.1575  0.1699  0.1844 -0.2933  0.0835
#> 
#> (2,.,.) = 
#>   0.2751 -0.3898  0.1213 -3.8687  0.4497
#>  -0.2111 -0.3950 -2.9382  3.1111  0.5782
#>  -0.1614  0.6500  0.7915  3.2278  0.2688
#> 
#> (3,.,.) = 
#>   0.0278  0.4387 -0.2164  2.4021  0.2127
#>  -0.6834 -0.2079 -0.3860  3.2684  1.3006
#>   0.1347 -0.8121 -0.8852 -3.9437  0.9420
#> 
#> (4,.,.) = 
#>   0.1821 -0.1450  1.3846 -3.6279 -0.6935
#>   0.0898  0.1431 -0.9339  1.3874  0.9015
#>   0.2645  1.0526  2.7827  0.2936 -1.0807
#> 
#> (5,.,.) = 
#>  -0.0958  0.2999  1.7004 -0.1108 -0.6775
#>   0.4078  0.8023  1.3688  0.0159 -0.5682
#>   0.1058  0.1554  1.2802 -1.5900  0.2783
#> 
#> (6,.,.) = 
#>  -0.1285  0.3316  2.4156 -0.7307 -0.6917
#>   0.2220  0.2726  0.4376 -0.4109 -0.4614
#>   0.1901 -0.4411  0.4385 -3.9330 -0.6751
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]