Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -3.7814 -0.7132 -0.3839 -0.1780  1.8651
#>   0.1242 -4.2034  1.5017  1.6880  4.2393
#>  -1.7262  0.2897 -0.7342 -2.9376  0.7744
#> 
#> (2,.,.) = 
#>  -0.3358  1.4806 -0.6356 -1.1463 -1.2927
#>   0.1363  1.4953 -0.4378 -1.6136 -1.2650
#>   3.4183  3.6155  0.6655  0.9884 -4.6952
#> 
#> (3,.,.) = 
#>   1.6643e+00  9.6982e-01  8.8248e-01  1.4876e+00 -1.5523e+00
#>  -1.7486e+00  8.8640e-01  2.8432e-01 -6.5280e-01 -8.6248e-05
#>   2.3860e+00 -9.7200e-01  7.7819e-01  7.1079e-01  2.7527e-01
#> 
#> (4,.,.) = 
#>   1.3233  2.5044 -0.0331  0.8960 -3.1524
#>   1.9993 -1.1982  1.3409  1.7782  0.5703
#>   1.8244  0.3172  0.0562 -0.7424 -0.7073
#> 
#> (5,.,.) = 
#>  -0.7531 -0.3605 -0.4185  0.2859  0.3666
#>   3.9144 -0.3923  2.1537  2.4777 -0.7262
#>  -1.5917  0.8062 -0.2800 -0.5097 -0.2348
#> 
#> (6,.,.) = 
#>   0.2916 -0.9768  0.7935  2.2200  0.5938
#>   0.4581 -1.6683 -0.5177 -1.6423  1.7106
#>   0.5185  2.1274 -0.2288 -0.5040 -2.2224
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]