Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.4642 -1.8618  2.9305  1.9857 -1.9997
#>   0.1906 -0.0002 -2.1750 -0.4812  1.3244
#>  -0.2329  0.5706  1.3064 -0.1734 -0.5518
#> 
#> (2,.,.) = 
#>  -3.1311  1.2975 -4.4469 -3.4690  3.7258
#>   0.3375 -0.0865 -0.7950 -0.0758 -0.0295
#>  -3.0946  0.7749 -1.7161 -2.3241  2.0543
#> 
#> (3,.,.) = 
#>   1.7237 -1.1199  4.4127  2.6827 -3.5945
#>   1.5970  0.5104 -1.5740 -0.1393  0.5831
#>   1.3288 -0.3249 -1.5754  0.4509  0.8856
#> 
#> (4,.,.) = 
#>   3.9646 -0.7174 -0.2573  2.2208 -0.5589
#>   4.3350  0.9775  1.1050  1.2451 -2.2934
#>   1.7890  1.5366 -1.3914 -0.6643  1.2944
#> 
#> (5,.,.) = 
#>   1.1548  2.4468 -1.4219 -1.8836  0.2393
#>   4.2307  1.9120 -2.1112 -0.2119  0.6237
#>   0.6762 -2.2889  0.3924  2.1799 -0.5128
#> 
#> (6,.,.) = 
#>  -0.9368 -1.6695  1.6256  1.4625 -0.1505
#>  -0.0059  0.1482  1.0101  0.1105 -0.8038
#>  -0.2148  1.8704 -0.5040 -1.8009 -0.1800
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]