Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -1.9045 -0.1634 -0.6373 -1.2935 -0.5695
#>  -1.1273 -2.2107 -3.8427 -5.7093  2.6768
#>   1.0917 -0.1711 -0.1448  0.2307  0.8366
#> 
#> (2,.,.) = 
#>   1.0714 -0.7180  0.3612 -0.6954 -2.6333
#>  -0.5768  1.3950  1.0388  2.9350  1.0109
#>   1.1250  0.0678  0.7938  2.2797 -2.1318
#> 
#> (3,.,.) = 
#>  -2.8325  0.5163  0.0211 -0.5058 -0.8806
#>  -1.9125  2.4282  1.3856  3.5557  2.9185
#>   1.8569 -1.0342  0.7042 -0.3541 -4.3848
#> 
#> (4,.,.) = 
#>  -0.4747  1.3056  1.8900  2.1825 -1.1278
#>   0.9909  1.2020  1.9748  1.2698  0.7608
#>   0.9964 -3.0578 -2.8205 -4.6357 -2.4018
#> 
#> (5,.,.) = 
#>  -0.8568  0.8809  0.6370  3.6000 -1.2871
#>  -2.1685 -0.0410 -1.3487 -0.0979  0.7874
#>   0.5708 -1.2393 -1.3833 -3.9753  1.5145
#> 
#> (6,.,.) = 
#>  -0.2355 -1.8067 -2.4437 -1.4694 -1.2864
#>  -0.4796 -1.1001 -1.3463 -3.1613  0.0874
#>   0.7157  0.0651  0.3916 -0.3660  0.4494
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]