Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -2.7853  2.1960  2.3967 -1.1556 -1.9306
#>   1.8081  2.1898 -3.0786 -1.2437  1.7806
#>   0.3912 -1.2600  0.4225  3.0386 -0.1885
#> 
#> (2,.,.) = 
#>   0.7105 -3.2254  0.1351 -0.2062  0.1635
#>  -3.3580  3.8511  4.5471  3.3653  0.7932
#>  -1.8094  0.5339  1.8887 -0.2150 -1.5302
#> 
#> (3,.,.) = 
#>  -1.2586  3.3245  0.0867 -2.3957 -0.2440
#>  -0.6557  1.7915  0.9232  0.1685  1.5621
#>  -1.6190 -2.8393  3.1001 -1.3783 -0.1415
#> 
#> (4,.,.) = 
#>  -1.7855 -0.2042  3.0917  2.0579 -0.1216
#>  -2.0324  3.3011  1.7920  0.9581 -0.5940
#>   2.0544 -3.6353 -1.2162  1.1380  0.7701
#> 
#> (5,.,.) = 
#>  -0.5468  1.9324 -0.4402 -3.3301  0.3384
#>   0.3017  2.8569 -1.1096  0.8356  0.5656
#>  -1.4124  2.3234  1.3164  2.3164 -1.1055
#> 
#> (6,.,.) = 
#>  -0.6782  0.1238  0.4772 -1.5948 -0.4430
#>  -0.1192 -3.8205  1.0234 -2.1605 -0.2842
#>  -1.4109 -0.3494  2.5632  1.1358  0.3197
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]