Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>   0.5548  2.3928  3.1336 -1.3894 -0.0623
#>  -0.7196  3.2582  7.3889  3.5047  4.0349
#>  -0.7642  0.5998 -1.5632  2.5094  1.5776
#> 
#> (2,.,.) = 
#>   1.2181  1.3334  2.8053  1.5557 -0.6960
#>  -1.1038  2.2033 -0.6700  0.1887  2.0798
#>  -0.0417 -0.1080 -2.6879 -0.5985 -0.6176
#> 
#> (3,.,.) = 
#>  -0.6592  0.9274 -0.9760  0.1862  1.0653
#>   0.1111 -0.1938 -0.0256  4.7286  0.8685
#>  -0.3843 -1.0181 -2.1289 -2.2640 -0.5548
#> 
#> (4,.,.) = 
#>   0.4200 -2.4388 -0.3099 -2.3074 -1.7514
#>   0.7028 -1.0933 -0.7764 -1.4621 -1.8006
#>  -0.0353 -2.7541 -3.4747 -1.2686 -1.4915
#> 
#> (5,.,.) = 
#>   1.5301 -1.7717 -2.3895 -0.3191 -3.2701
#>   0.3127  1.2952  5.1547  3.0431  1.4898
#>   0.5684 -1.3643 -3.4956 -3.1960 -2.5763
#> 
#> (6,.,.) = 
#>  -0.2708  0.6337  0.9569  1.8260  1.1560
#>  -0.3799  3.6413  5.2594  6.3147  3.8236
#>   0.7433 -1.6495 -1.9338  0.5066 -1.7528
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]