Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   1.9935  1.7517  0.8616  0.4387  0.2436
#>  -4.4862 -3.1653  5.5999  0.9638 -4.1026
#>  -2.3447 -1.0177 -2.0951 -0.4424  1.6604
#> 
#> (2,.,.) = 
#>   1.6303  2.4312  4.6162 -0.6833 -1.7308
#>   1.3296 -4.0179 -3.4181  2.4787 -2.1685
#>   1.7412  1.6944  1.8478 -0.4883 -0.6726
#> 
#> (3,.,.) = 
#>   1.7651  1.2606 -2.4064 -0.6104  1.6848
#>  -5.1041 -0.6864 -1.3065 -1.8128  2.8055
#>  -2.0941 -2.1288 -0.2080 -0.7060 -1.1500
#> 
#> (4,.,.) = 
#>  -3.9267 -1.3144 -1.7412 -1.6936  1.6611
#>   1.2031  2.1622  1.2810  0.0450  0.7961
#>   0.9636  0.5091 -1.4130  0.6447  1.1467
#> 
#> (5,.,.) = 
#>  -3.6345 -3.5363 -1.8716  0.2400 -0.2416
#>  -1.2450 -0.3076  1.6796 -0.7329 -0.9484
#>  -0.4651  0.3524 -0.2125 -0.4818  0.6720
#> 
#> (6,.,.) = 
#>  -2.6872 -4.6670 -0.9108  2.0336 -2.1385
#>   0.7031 -0.7976 -1.6803  0.3957 -0.0702
#>   1.3601 -0.7551  1.0828  1.5511 -1.9165
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]