Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -1.3186 -0.2586 -0.7057  0.6547  0.0069
#>  -4.2472  5.3279  4.0181  3.5163 -0.1500
#>  -0.7165 -0.4003 -0.5469  0.2776 -0.2834
#> 
#> (2,.,.) = 
#>   1.8586 -0.4130 -0.0016 -1.0608 -0.2725
#>  -4.2259  4.7033  2.1293  3.6022  0.3636
#>  -0.7472  0.6876  1.8026  0.3038 -0.5717
#> 
#> (3,.,.) = 
#>  -0.1676  0.3717 -2.5723  0.7373 -2.2723
#>   0.5899  0.2354 -0.8862 -0.0135  0.7921
#>  -0.9955  0.4053  0.9501  0.4675 -1.1173
#> 
#> (4,.,.) = 
#>   4.4547 -5.6439 -6.5889 -3.2580  0.8657
#>   0.5844 -1.1059  0.2961 -0.8036 -0.4800
#>  -2.9052  3.0637  1.8151  2.3391 -0.5539
#> 
#> (5,.,.) = 
#>  -0.3130  0.9538  1.1233  0.3408 -0.8832
#>   2.5876  0.5535 -0.2109 -0.9668  2.9441
#>   0.6168  2.3188  0.5862  0.5739  1.2918
#> 
#> (6,.,.) = 
#>  -4.0559  5.5334  1.6362  3.9690 -0.0127
#>   0.0342  2.3361  4.3022  0.1444  1.9093
#>   2.5099 -4.2258 -2.8709 -2.4392 -0.1778
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]