Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.5084 -1.1742  0.8821  0.1727 -2.4295
#>  -1.4485 -1.5814 -0.7322  2.9576  2.7180
#>  -2.1679  4.0392  0.0087  2.0460  0.5875
#> 
#> (2,.,.) = 
#>  -0.9811 -5.2635  0.4626  2.0430  1.2745
#>   2.1923 -0.0387 -0.3332 -2.6556 -0.7344
#>   3.9212 -1.6830  0.5486 -5.4397 -2.5903
#> 
#> (3,.,.) = 
#>  -3.1654  2.4020  0.8836  3.0109 -2.7384
#>   2.0150 -1.9208 -0.6268 -1.7655  1.4853
#>  -1.4320  4.0295  0.5593  0.5479 -0.3580
#> 
#> (4,.,.) = 
#>   1.6694  1.3480  1.2561 -3.6998 -2.5497
#>  -0.8046 -0.2175  0.3342  0.8904 -3.1455
#>   0.3192 -1.5151  0.5671 -0.6178 -0.7057
#> 
#> (5,.,.) = 
#>  -4.1313  4.5848  2.2339  2.6183 -4.5381
#>  -0.6289 -1.2707 -1.2003  2.1690  4.8901
#>  -3.7929 -1.0057 -0.7198  6.0015  3.4721
#> 
#> (6,.,.) = 
#>   0.9356 -0.4470 -1.2361 -0.0776  3.0080
#>   3.7697  1.4845 -0.4456 -5.0931  2.3091
#>   0.2893 -2.4663 -1.5024  1.5103  2.3114
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]