Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor 
#> (1,.,.) = 
#>  -4.3965 -1.2687  1.0527 -1.3582 -2.0137
#>   1.7526  0.4774 -1.2786 -1.2011  0.3185
#>  -0.7419  0.3712 -1.1304 -0.6134 -2.1909
#> 
#> (2,.,.) = 
#>   1.0328  0.6238 -0.7075  0.5685  0.2156
#>  -2.3869  0.3818 -1.2121 -0.5598 -0.5324
#>  -1.1057 -0.5324  0.3456 -0.9070 -0.7164
#> 
#> (3,.,.) = 
#>   0.8790  0.7950 -1.6822 -0.7317  0.3915
#>  -1.7487  0.1048  0.4513  1.6684 -1.1785
#>  -2.2020 -0.8638  0.7228 -1.2682  1.2481
#> 
#> (4,.,.) = 
#>   1.1382  0.3859 -0.4212  0.5012 -2.5275
#>   1.9565 -0.2553  2.1054  2.7793  1.1618
#>  -0.4515  0.7311 -1.5641 -0.1970 -3.0944
#> 
#> (5,.,.) = 
#>  -3.0063 -0.4141 -0.2822 -1.2631 -2.0668
#>   0.7015  0.9527  0.5719  4.3668 -0.9080
#>  -0.7993 -0.7693  2.2251  1.9926 -1.9487
#> 
#> (6,.,.) = 
#>   1.2756 -0.0823  0.5286  0.3705  1.7703
#>  -0.8725 -0.2535  1.8573  3.0133 -1.1599
#>   1.5309  0.5303 -2.6080 -3.7462  2.4240
#> 
#> (7,.,.) = 
#>  -1.3704 -0.4300  2.5480  3.8088 -0.6782
#>   0.8254  0.3362 -0.4954  0.1036 -0.7812
#>  -2.1904 -1.3008  2.2895  0.4921 -1.4129
#> 
#> (8,.,.) = 
#>  -2.3587  0.0766 -0.2034  0.2611  0.6048
#>   1.5406  1.6147 -4.1611 -2.9916  1.9736
#>   4.1022  1.2071 -1.6255  0.1183  1.4958
#> 
#> (9,.,.) = 
#>   1.9838  0.0539  0.0786 -0.1013  0.8964
#>  -0.0521 -0.1591  0.4983  0.5100 -0.9648
#>  -1.3628 -0.7750  1.4627  0.4237  0.0907
#> 
#> (10,.,.) = 
#>  -0.5810  0.0472  0.9418  2.3085 -2.4165
#>   1.1348  0.1767 -0.8898 -1.4204  0.9883
#>   1.3788  0.0744  0.3734  0.8719 -1.7745
#> [ CPUFloatType{10,3,5} ]