Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -1.5116 -2.7277 -6.4360 -0.2134  1.3064
#>   0.0625 -1.4222 -0.4484  1.2628  0.5458
#>  -1.4283 -3.8036 -6.2693  1.0423  2.1332
#> 
#> (2,.,.) = 
#>  -0.4159  0.3043 -0.4568 -0.3449  0.1111
#>  -3.8927  0.2111 -5.7974 -1.1443  0.1748
#>   4.2864 -0.1018  4.2249 -0.6210  2.3224
#> 
#> (3,.,.) = 
#>  -1.0395 -1.1628 -4.0869 -0.7082  0.4061
#>  -1.6418 -6.0883 -8.9015  1.8672  4.3814
#>  -4.9950 -2.1603 -6.3411  2.1921  0.0556
#> 
#> (4,.,.) = 
#>  -3.1311 -1.9400 -3.3251  2.6556 -1.6895
#>   0.5605  0.1842  0.6647 -0.1677 -0.5920
#>  -2.3682  1.4583 -3.3031 -2.0942 -0.0125
#> 
#> (5,.,.) = 
#>   0.9081  0.4767  2.1242  0.1634  0.4314
#>  -2.9560  0.3270 -2.6540  0.2644 -0.9712
#>  -0.6639  1.6508 -1.0901 -2.0587 -1.4760
#> 
#> (6,.,.) = 
#>  -1.7207  2.5228 -0.5519 -1.9306  0.0525
#>   0.5605 -1.8737  0.2245  1.8262  0.9840
#>   0.9720 -1.6250 -1.1108  0.1393  2.2859
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]