Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   2.0321 -3.4536 -4.2306 -0.7345 -2.8552
#>   0.2361 -2.1475 -1.4500 -0.5769 -1.0370
#>  -1.2193 -3.9016  0.3903 -1.0224 -0.4516
#> 
#> (2,.,.) = 
#>  -0.2836  4.4929  0.8841 -0.2261  1.1850
#>   0.6023  2.4837  2.8513 -1.3269  0.0882
#>  -3.0624  0.9875 -1.7773  0.9040  2.1253
#> 
#> (3,.,.) = 
#>  -0.4699 -0.5460 -1.2779 -2.4299 -0.9504
#>  -0.1150  1.6523 -0.7865  1.7993  0.9980
#>  -1.5145 -0.2638  3.6538 -0.7750  1.0817
#> 
#> (4,.,.) = 
#>   1.2806 -0.3951  0.0297  1.5808 -0.2420
#>   1.3521 -0.6699 -3.0296 -0.8832 -1.7286
#>  -1.1154 -1.0209  1.5944  2.3739  1.5806
#> 
#> (5,.,.) = 
#>   2.8573  2.6045 -1.2535 -4.1867 -2.9171
#>  -3.2812 -3.7681 -1.0245  3.3926  2.2834
#>  -0.2319 -1.1268 -3.4610  0.4558 -0.4358
#> 
#> (6,.,.) = 
#>   1.4121  0.2034 -3.0405 -1.9328 -1.9768
#>   0.5806 -1.3909  1.7354  0.4146 -0.2369
#>  -0.9771 -0.7393  1.7604 -0.6632  0.4255
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]