Matmul

Matmul

torch_matmul(self, other)

Arguments

self
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}

#> torch_tensor
#> (1,.,.) = 
#>  -0.3754  1.4558  1.8240  0.0396 -2.3706
#>   5.0175 -0.5841 -0.4681 -2.9149 -0.3252
#>   3.2460 -3.1777  1.2949 -2.6804 -0.2619
#> 
#> (2,.,.) = 
#>  -3.2304  2.4509  1.6269  2.9341 -2.7387
#>  -2.3756  4.8950 -2.0618  2.3903  0.0496
#>   1.7730 -1.2391  2.3687 -1.0092 -2.4172
#> 
#> (3,.,.) = 
#>   4.1876 -3.9260  0.1832 -2.7171  0.7214
#>   0.8378  0.8429 -0.1005 -1.1376 -0.0135
#>   3.3547  0.4235 -2.8978  0.3162  0.7977
#> 
#> (4,.,.) = 
#>   1.9757 -0.7985 -1.7246  0.4994  0.7287
#>  -2.0489  4.3341 -2.5646  2.9921  0.2565
#>  -1.8314  1.4251 -1.8264  1.9419  1.2071
#> 
#> (5,.,.) = 
#>  -6.7118 -3.6599  1.5914  0.5205  3.3643
#>   2.1202 -4.0287 -0.4667 -2.6594  2.6191
#>  -2.8975 -5.0203  7.0576 -1.6417 -2.7818
#> 
#> (6,.,.) = 
#>   1.4011 -2.6312  1.3619 -2.2871  0.2577
#>   2.7245  0.2498 -2.8973 -0.6318  1.8274
#>  -2.1711  1.4851 -0.5742  1.5788  0.2624
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]