Matmul

Matmul

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.1504 -1.3041 -2.5564 -0.2499  0.9351
#>  -4.0406  1.7200  1.2851  1.9503 -1.2225
#>   3.7554  0.3982  3.3424  0.3714 -2.1479
#> 
#> (2,.,.) = 
#>  -0.6528 -1.2814 -1.5917 -0.9285  1.7415
#>  -0.7955 -2.4634  0.2554  5.4232 -4.6318
#>   3.4262 -0.5508  1.4093 -3.4563  2.4334
#> 
#> (3,.,.) = 
#>  -6.0321  0.7785 -2.8900 -0.7499  2.9618
#>   1.4079  2.4440  1.6150 -3.7219  2.4830
#>  -2.5101  2.4646 -0.3251 -1.2679  1.4870
#> 
#> (4,.,.) = 
#>  -5.8172  0.7193 -3.1951  1.2748  0.8087
#>  -2.1722  0.8543 -0.7270 -1.9478  2.6159
#>  -2.9006 -0.1775 -1.1778 -1.5731  2.8495
#> 
#> (5,.,.) = 
#>   1.3571  0.4367  3.1463 -0.0706 -0.8121
#>  -2.4099  0.3029 -0.6517  0.1927  0.6296
#>   0.8097  0.7823  4.2283  1.5186 -2.4737
#> 
#> (6,.,.) = 
#>  -1.1114  0.5995 -1.2152 -1.6292  2.0181
#>  -1.0162 -0.8917 -0.0609  1.2364 -0.6397
#>  -0.7436 -0.6224 -1.9510  1.8321 -1.2925
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]