Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   4.7803  3.0229  2.1551  2.8125  2.7314
#>   0.7068 -2.7891 -1.5522 -0.2826 -0.4970
#>   4.2077 -1.6328  0.5911  1.6349 -0.4065
#> 
#> (2,.,.) = 
#>  -4.5370  2.7957  0.0816 -0.8152  1.1204
#>   4.7366 -1.2363  0.4410 -0.2294 -1.1916
#>  -0.6317  4.4316  1.3404  0.0006  2.1301
#> 
#> (3,.,.) = 
#>  -3.4745 -2.5872 -2.4263 -3.1506 -1.9857
#>  -1.7505 -0.0412 -0.2544 -0.3059 -0.2009
#>  -0.7211 -2.2949 -0.9488 -0.1848 -1.0549
#> 
#> (4,.,.) = 
#>   2.5844 -1.8491 -0.1807 -0.9281 -1.7753
#>  -1.6396  1.2921 -0.3241 -1.6686  0.1002
#>   4.6138 -0.4967  1.4569  3.1472  0.6682
#> 
#> (5,.,.) = 
#>   1.4547  0.2730  0.5717  2.0917  1.3120
#>   2.5752 -0.0575  0.5506  1.6882  0.9008
#>   1.3289  2.3007  1.1468  1.7932  2.1345
#> 
#> (6,.,.) = 
#>   3.3893  0.9074  1.6295  2.3851  0.9398
#>  -5.1480 -0.7659 -2.1418 -4.0713 -1.6812
#>   0.3525 -1.5174 -0.8846 -0.6901 -0.6377
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]