Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>  -0.5083  1.1115  0.8702 -1.3267 -0.5502
#>  -0.1910  0.0157 -0.2893  0.7561 -0.2460
#>   0.7410  0.6425 -0.1564 -0.3902 -0.5873
#> 
#> (2,.,.) = 
#>   5.9672 -4.3506 -1.7180  0.6210  1.0916
#>   2.3804 -3.3517 -0.7981  0.8189  1.5617
#>   3.3729 -2.5022 -1.5039  0.3391  1.5729
#> 
#> (3,.,.) = 
#>   3.9675 -3.6310 -1.5842 -0.7152  3.5367
#>  -3.6736  2.3443  1.6079 -0.1058 -1.5936
#>   0.9141 -0.1077  0.2638 -0.5335 -0.5370
#> 
#> (4,.,.) = 
#>  -0.4071  0.7285  0.0768  0.1489 -0.6618
#>  -2.7210  1.1955  1.3492 -0.9597  0.1305
#>  -2.9640  0.1111 -1.1135  1.3119  2.9070
#> 
#> (5,.,.) = 
#>   1.7666 -1.5223  0.3325  0.6024 -1.3560
#>  -2.9904  3.0255  0.0817  0.1270 -0.5909
#>  -3.1859  3.8498  0.6708 -0.3277 -1.6924
#> 
#> (6,.,.) = 
#>  -0.6592 -0.8033 -0.6149 -0.0245  2.4636
#>  -0.9089  1.3087  2.9317 -1.9489 -3.1277
#>   2.8054 -3.8199 -1.9769  2.0512  2.1786
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]