Matmul

Matmul

Usage

torch_matmul(self, other)

Arguments

self

(Tensor) the first tensor to be multiplied

other

(Tensor) the second tensor to be multiplied

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=NULL) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

if (torch_is_installed()) {

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)
# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)
# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)
# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)
}
#> torch_tensor
#> (1,.,.) = 
#>   0.4346 -1.4957 -0.0119 -1.5183 -1.3012
#>  -0.9312  1.0402  0.3585  3.6638 -0.8190
#>   0.2518  0.4582 -0.1414  0.5834  0.1219
#> 
#> (2,.,.) = 
#>  -0.8226  1.1543  0.3918  2.1950  0.5286
#>  -3.1902  1.1740  1.2736  1.9057  1.9050
#>  -3.3993  2.6058  1.6024  1.8606  4.4739
#> 
#> (3,.,.) = 
#>   0.8728  0.4272 -1.1861 -0.2314 -0.0270
#>  -1.9020  1.0872  1.4273  1.7817  1.6167
#>   0.4717  1.0696 -1.0483  2.1772 -0.7420
#> 
#> (4,.,.) = 
#>  -1.0891 -0.1733  0.9948 -1.4676  1.7404
#>  -0.3341  2.2035 -0.2571  3.9021  0.3508
#>  -1.0984  0.6770  0.0860  0.6736  0.8800
#> 
#> (5,.,.) = 
#>   0.4992  1.4951 -0.3096  4.2879 -1.3386
#>  -1.6613 -0.3617  1.6366 -2.0699  2.4312
#>   2.7820 -2.4281 -1.8558 -3.1242 -3.1612
#> 
#> (6,.,.) = 
#>  -0.8288 -0.5607  0.1310 -2.3068  1.2360
#>   4.0450 -4.9250 -2.1532 -5.1119 -6.0670
#>   4.3940 -1.4290 -3.1910 -0.4135 -4.9012
#> 
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,3,5} ]